Network.AWS.Redshift.ModifyClusterIamRoles

Description

Modifies the list of Identity and Access Management (IAM) roles that can be used by the cluster to access other Amazon Web Services services.

A cluster can have up to 10 IAM roles associated at any time.

Creating a Request

data ModifyClusterIamRoles

See: newModifyClusterIamRoles smart constructor.

Constructors

ModifyClusterIamRoles'

Fields

removeIamRoles :: Maybe [Text]

Zero or more IAM roles in ARN format to disassociate from the cluster. You can disassociate up to 10 IAM roles from a single cluster in a single request.

addIamRoles :: Maybe [Text]

Zero or more IAM roles to associate with the cluster. The roles must be in their Amazon Resource Name (ARN) format. You can associate up to 10 IAM roles with a single cluster in a single request.

clusterIdentifier :: Text

The unique identifier of the cluster for which you want to associate or disassociate IAM roles.

newModifyClusterIamRoles

Arguments

:: Text
$sel:clusterIdentifier:ModifyClusterIamRoles'

-> ModifyClusterIamRoles

Create a value of ModifyClusterIamRoles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:removeIamRoles:ModifyClusterIamRoles', modifyClusterIamRoles_removeIamRoles - Zero or more IAM roles in ARN format to disassociate from the cluster. You can disassociate up to 10 IAM roles from a single cluster in a single request.

$sel:addIamRoles:ModifyClusterIamRoles', modifyClusterIamRoles_addIamRoles - Zero or more IAM roles to associate with the cluster. The roles must be in their Amazon Resource Name (ARN) format. You can associate up to 10 IAM roles with a single cluster in a single request.

$sel:clusterIdentifier:ModifyClusterIamRoles', modifyClusterIamRoles_clusterIdentifier - The unique identifier of the cluster for which you want to associate or disassociate IAM roles.

Request Lenses

modifyClusterIamRoles_removeIamRoles :: Lens' ModifyClusterIamRoles (Maybe [Text])

Zero or more IAM roles in ARN format to disassociate from the cluster. You can disassociate up to 10 IAM roles from a single cluster in a single request.

modifyClusterIamRoles_addIamRoles :: Lens' ModifyClusterIamRoles (Maybe [Text])

Zero or more IAM roles to associate with the cluster. The roles must be in their Amazon Resource Name (ARN) format. You can associate up to 10 IAM roles with a single cluster in a single request.

modifyClusterIamRoles_clusterIdentifier :: Lens' ModifyClusterIamRoles Text

The unique identifier of the cluster for which you want to associate or disassociate IAM roles.

Destructuring the Response

data ModifyClusterIamRolesResponse

See: newModifyClusterIamRolesResponse smart constructor.

Constructors

ModifyClusterIamRolesResponse'

Fields

cluster :: Maybe Cluster

httpStatus :: Int

The response's http status code.

newModifyClusterIamRolesResponse

Arguments

:: Int
$sel:httpStatus:ModifyClusterIamRolesResponse'

-> ModifyClusterIamRolesResponse

Create a value of ModifyClusterIamRolesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:ModifyClusterIamRolesResponse', modifyClusterIamRolesResponse_cluster - Undocumented member.

$sel:httpStatus:ModifyClusterIamRolesResponse', modifyClusterIamRolesResponse_httpStatus - The response's http status code.

Response Lenses

modifyClusterIamRolesResponse_cluster :: Lens' ModifyClusterIamRolesResponse (Maybe Cluster)

Undocumented member.

modifyClusterIamRolesResponse_httpStatus :: Lens' ModifyClusterIamRolesResponse Int

The response's http status code.