Network.AWS.Redshift.ModifyClusterMaintenance

Description

Modifies the maintenance settings of a cluster.

Creating a Request

data ModifyClusterMaintenance

See: newModifyClusterMaintenance smart constructor.

Constructors

ModifyClusterMaintenance'

Fields

deferMaintenanceIdentifier :: Maybe Text

A unique identifier for the deferred maintenance window.

deferMaintenanceDuration :: Maybe Int

An integer indicating the duration of the maintenance window in days. If you specify a duration, you can't specify an end time. The duration must be 45 days or less.

deferMaintenanceStartTime :: Maybe ISO8601

A timestamp indicating the start time for the deferred maintenance window.

deferMaintenanceEndTime :: Maybe ISO8601

A timestamp indicating end time for the deferred maintenance window. If you specify an end time, you can't specify a duration.

deferMaintenance :: Maybe Bool

A boolean indicating whether to enable the deferred maintenance window.

clusterIdentifier :: Text

A unique identifier for the cluster.

newModifyClusterMaintenance

Arguments

:: Text
$sel:clusterIdentifier:ModifyClusterMaintenance'

-> ModifyClusterMaintenance

Create a value of ModifyClusterMaintenance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deferMaintenanceIdentifier:ModifyClusterMaintenance', modifyClusterMaintenance_deferMaintenanceIdentifier - A unique identifier for the deferred maintenance window.

$sel:deferMaintenanceDuration:ModifyClusterMaintenance', modifyClusterMaintenance_deferMaintenanceDuration - An integer indicating the duration of the maintenance window in days. If you specify a duration, you can't specify an end time. The duration must be 45 days or less.

$sel:deferMaintenanceStartTime:ModifyClusterMaintenance', modifyClusterMaintenance_deferMaintenanceStartTime - A timestamp indicating the start time for the deferred maintenance window.

$sel:deferMaintenanceEndTime:ModifyClusterMaintenance', modifyClusterMaintenance_deferMaintenanceEndTime - A timestamp indicating end time for the deferred maintenance window. If you specify an end time, you can't specify a duration.

$sel:deferMaintenance:ModifyClusterMaintenance', modifyClusterMaintenance_deferMaintenance - A boolean indicating whether to enable the deferred maintenance window.

$sel:clusterIdentifier:ModifyClusterMaintenance', modifyClusterMaintenance_clusterIdentifier - A unique identifier for the cluster.

Request Lenses

modifyClusterMaintenance_deferMaintenanceIdentifier :: Lens' ModifyClusterMaintenance (Maybe Text)

A unique identifier for the deferred maintenance window.

modifyClusterMaintenance_deferMaintenanceDuration :: Lens' ModifyClusterMaintenance (Maybe Int)

An integer indicating the duration of the maintenance window in days. If you specify a duration, you can't specify an end time. The duration must be 45 days or less.

modifyClusterMaintenance_deferMaintenanceStartTime :: Lens' ModifyClusterMaintenance (Maybe UTCTime)

A timestamp indicating the start time for the deferred maintenance window.

modifyClusterMaintenance_deferMaintenanceEndTime :: Lens' ModifyClusterMaintenance (Maybe UTCTime)

A timestamp indicating end time for the deferred maintenance window. If you specify an end time, you can't specify a duration.

modifyClusterMaintenance_deferMaintenance :: Lens' ModifyClusterMaintenance (Maybe Bool)

A boolean indicating whether to enable the deferred maintenance window.

modifyClusterMaintenance_clusterIdentifier :: Lens' ModifyClusterMaintenance Text

A unique identifier for the cluster.

Destructuring the Response

data ModifyClusterMaintenanceResponse

See: newModifyClusterMaintenanceResponse smart constructor.

Constructors

ModifyClusterMaintenanceResponse'

Fields

cluster :: Maybe Cluster

httpStatus :: Int

The response's http status code.

newModifyClusterMaintenanceResponse

Arguments

:: Int
$sel:httpStatus:ModifyClusterMaintenanceResponse'

-> ModifyClusterMaintenanceResponse

Create a value of ModifyClusterMaintenanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:ModifyClusterMaintenanceResponse', modifyClusterMaintenanceResponse_cluster - Undocumented member.

$sel:httpStatus:ModifyClusterMaintenanceResponse', modifyClusterMaintenanceResponse_httpStatus - The response's http status code.

Response Lenses

modifyClusterMaintenanceResponse_cluster :: Lens' ModifyClusterMaintenanceResponse (Maybe Cluster)

Undocumented member.

modifyClusterMaintenanceResponse_httpStatus :: Lens' ModifyClusterMaintenanceResponse Int

The response's http status code.