Network.AWS.Redshift.ModifyEndpointAccess

Description

Modifies a Redshift-managed VPC endpoint.

Creating a Request

data ModifyEndpointAccess

See: newModifyEndpointAccess smart constructor.

Constructors

ModifyEndpointAccess'

Fields

vpcSecurityGroupIds :: Maybe [Text]

The complete list of VPC security groups associated with the endpoint after the endpoint is modified.

endpointName :: Text

The endpoint to be modified.

newModifyEndpointAccess

Arguments

:: Text
$sel:endpointName:ModifyEndpointAccess'

-> ModifyEndpointAccess

Create a value of ModifyEndpointAccess with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcSecurityGroupIds:ModifyEndpointAccess', modifyEndpointAccess_vpcSecurityGroupIds - The complete list of VPC security groups associated with the endpoint after the endpoint is modified.

$sel:endpointName:ModifyEndpointAccess', modifyEndpointAccess_endpointName - The endpoint to be modified.

Request Lenses

modifyEndpointAccess_vpcSecurityGroupIds :: Lens' ModifyEndpointAccess (Maybe [Text])

The complete list of VPC security groups associated with the endpoint after the endpoint is modified.

modifyEndpointAccess_endpointName :: Lens' ModifyEndpointAccess Text

The endpoint to be modified.

Destructuring the Response

data EndpointAccess

Describes a Redshift-managed VPC endpoint.

See: newEndpointAccess smart constructor.

Constructors

EndpointAccess'

Fields

vpcSecurityGroups :: Maybe [VpcSecurityGroupMembership]

The security groups associated with the endpoint.

endpointName :: Maybe Text

The name of the endpoint.

address :: Maybe Text

The DNS address of the endpoint.

resourceOwner :: Maybe Text

The Amazon Web Services account ID of the owner of the cluster.

endpointCreateTime :: Maybe ISO8601

The time (UTC) that the endpoint was created.

endpointStatus :: Maybe Text

The status of the endpoint.

clusterIdentifier :: Maybe Text

The cluster identifier of the cluster associated with the endpoint.

port :: Maybe Int

The port number on which the cluster accepts incoming connections.

vpcEndpoint :: Maybe VpcEndpoint

subnetGroupName :: Maybe Text

The subnet group name where Amazon Redshift chooses to deploy the endpoint.

newEndpointAccess :: EndpointAccess

Create a value of EndpointAccess with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcSecurityGroups:EndpointAccess', endpointAccess_vpcSecurityGroups - The security groups associated with the endpoint.

$sel:endpointName:EndpointAccess', endpointAccess_endpointName - The name of the endpoint.

$sel:address:EndpointAccess', endpointAccess_address - The DNS address of the endpoint.

$sel:resourceOwner:EndpointAccess', endpointAccess_resourceOwner - The Amazon Web Services account ID of the owner of the cluster.

$sel:endpointCreateTime:EndpointAccess', endpointAccess_endpointCreateTime - The time (UTC) that the endpoint was created.

$sel:endpointStatus:EndpointAccess', endpointAccess_endpointStatus - The status of the endpoint.

$sel:clusterIdentifier:EndpointAccess', endpointAccess_clusterIdentifier - The cluster identifier of the cluster associated with the endpoint.

$sel:port:EndpointAccess', endpointAccess_port - The port number on which the cluster accepts incoming connections.

$sel:vpcEndpoint:EndpointAccess', endpointAccess_vpcEndpoint - Undocumented member.

$sel:subnetGroupName:EndpointAccess', endpointAccess_subnetGroupName - The subnet group name where Amazon Redshift chooses to deploy the endpoint.

Response Lenses

endpointAccess_vpcSecurityGroups :: Lens' EndpointAccess (Maybe [VpcSecurityGroupMembership])

The security groups associated with the endpoint.

endpointAccess_endpointName :: Lens' EndpointAccess (Maybe Text)

The name of the endpoint.

endpointAccess_address :: Lens' EndpointAccess (Maybe Text)

The DNS address of the endpoint.

endpointAccess_resourceOwner :: Lens' EndpointAccess (Maybe Text)

The Amazon Web Services account ID of the owner of the cluster.

endpointAccess_endpointCreateTime :: Lens' EndpointAccess (Maybe UTCTime)

The time (UTC) that the endpoint was created.

endpointAccess_endpointStatus :: Lens' EndpointAccess (Maybe Text)

The status of the endpoint.

endpointAccess_clusterIdentifier :: Lens' EndpointAccess (Maybe Text)

The cluster identifier of the cluster associated with the endpoint.

endpointAccess_port :: Lens' EndpointAccess (Maybe Int)

The port number on which the cluster accepts incoming connections.

endpointAccess_vpcEndpoint :: Lens' EndpointAccess (Maybe VpcEndpoint)

Undocumented member.

endpointAccess_subnetGroupName :: Lens' EndpointAccess (Maybe Text)

The subnet group name where Amazon Redshift chooses to deploy the endpoint.