Network.AWS.Redshift.ModifySnapshotSchedule

Description

Modifies a snapshot schedule. Any schedule associated with a cluster is modified asynchronously.

Creating a Request

data ModifySnapshotSchedule

See: newModifySnapshotSchedule smart constructor.

Constructors

ModifySnapshotSchedule'

Fields

scheduleIdentifier :: Text

A unique alphanumeric identifier of the schedule to modify.

scheduleDefinitions :: [Text]

An updated list of schedule definitions. A schedule definition is made up of schedule expressions, for example, "cron(30 12 *)" or "rate(12 hours)".

newModifySnapshotSchedule

Arguments

:: Text
$sel:scheduleIdentifier:ModifySnapshotSchedule'

-> ModifySnapshotSchedule

Create a value of ModifySnapshotSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scheduleIdentifier:ModifySnapshotSchedule', modifySnapshotSchedule_scheduleIdentifier - A unique alphanumeric identifier of the schedule to modify.

$sel:scheduleDefinitions:ModifySnapshotSchedule', modifySnapshotSchedule_scheduleDefinitions - An updated list of schedule definitions. A schedule definition is made up of schedule expressions, for example, "cron(30 12 *)" or "rate(12 hours)".

Request Lenses

modifySnapshotSchedule_scheduleIdentifier :: Lens' ModifySnapshotSchedule Text

A unique alphanumeric identifier of the schedule to modify.

modifySnapshotSchedule_scheduleDefinitions :: Lens' ModifySnapshotSchedule [Text]

An updated list of schedule definitions. A schedule definition is made up of schedule expressions, for example, "cron(30 12 *)" or "rate(12 hours)".

Destructuring the Response

data SnapshotSchedule

Describes a snapshot schedule. You can set a regular interval for creating snapshots of a cluster. You can also schedule snapshots for specific dates.

See: newSnapshotSchedule smart constructor.

Constructors

SnapshotSchedule'

Fields

nextInvocations :: Maybe [ISO8601]

associatedClusters :: Maybe [ClusterAssociatedToSchedule]

A list of clusters associated with the schedule. A maximum of 100 clusters is returned.

scheduleIdentifier :: Maybe Text

A unique identifier for the schedule.

scheduleDescription :: Maybe Text

The description of the schedule.

scheduleDefinitions :: Maybe [Text]

A list of ScheduleDefinitions.

tags :: Maybe [Tag]

An optional set of tags describing the schedule.

associatedClusterCount :: Maybe Int

The number of clusters associated with the schedule.

newSnapshotSchedule :: SnapshotSchedule

Create a value of SnapshotSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextInvocations:SnapshotSchedule', snapshotSchedule_nextInvocations -

$sel:associatedClusters:SnapshotSchedule', snapshotSchedule_associatedClusters - A list of clusters associated with the schedule. A maximum of 100 clusters is returned.

$sel:scheduleIdentifier:SnapshotSchedule', snapshotSchedule_scheduleIdentifier - A unique identifier for the schedule.

$sel:scheduleDescription:SnapshotSchedule', snapshotSchedule_scheduleDescription - The description of the schedule.

$sel:scheduleDefinitions:SnapshotSchedule', snapshotSchedule_scheduleDefinitions - A list of ScheduleDefinitions.

$sel:tags:SnapshotSchedule', snapshotSchedule_tags - An optional set of tags describing the schedule.

$sel:associatedClusterCount:SnapshotSchedule', snapshotSchedule_associatedClusterCount - The number of clusters associated with the schedule.

Response Lenses

snapshotSchedule_nextInvocations :: Lens' SnapshotSchedule (Maybe [UTCTime])

snapshotSchedule_associatedClusters :: Lens' SnapshotSchedule (Maybe [ClusterAssociatedToSchedule])

A list of clusters associated with the schedule. A maximum of 100 clusters is returned.

snapshotSchedule_scheduleIdentifier :: Lens' SnapshotSchedule (Maybe Text)

A unique identifier for the schedule.

snapshotSchedule_scheduleDescription :: Lens' SnapshotSchedule (Maybe Text)

The description of the schedule.

snapshotSchedule_scheduleDefinitions :: Lens' SnapshotSchedule (Maybe [Text])

A list of ScheduleDefinitions.

snapshotSchedule_tags :: Lens' SnapshotSchedule (Maybe [Tag])

An optional set of tags describing the schedule.

snapshotSchedule_associatedClusterCount :: Lens' SnapshotSchedule (Maybe Int)

The number of clusters associated with the schedule.