Network.AWS.Redshift.ModifyUsageLimit

Description

Modifies a usage limit in a cluster. You can't modify the feature type or period of a usage limit.

Creating a Request

data ModifyUsageLimit

See: newModifyUsageLimit smart constructor.

Constructors

ModifyUsageLimit'

Fields

amount :: Maybe Integer

The new limit amount. For more information about this parameter, see UsageLimit.

breachAction :: Maybe UsageLimitBreachAction

The new action that Amazon Redshift takes when the limit is reached. For more information about this parameter, see UsageLimit.

usageLimitId :: Text

The identifier of the usage limit to modify.

newModifyUsageLimit

Arguments

:: Text
$sel:usageLimitId:ModifyUsageLimit'

-> ModifyUsageLimit

Create a value of ModifyUsageLimit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:amount:ModifyUsageLimit', modifyUsageLimit_amount - The new limit amount. For more information about this parameter, see UsageLimit.

$sel:breachAction:ModifyUsageLimit', modifyUsageLimit_breachAction - The new action that Amazon Redshift takes when the limit is reached. For more information about this parameter, see UsageLimit.

$sel:usageLimitId:ModifyUsageLimit', modifyUsageLimit_usageLimitId - The identifier of the usage limit to modify.

Request Lenses

modifyUsageLimit_amount :: Lens' ModifyUsageLimit (Maybe Integer)

The new limit amount. For more information about this parameter, see UsageLimit.

modifyUsageLimit_breachAction :: Lens' ModifyUsageLimit (Maybe UsageLimitBreachAction)

The new action that Amazon Redshift takes when the limit is reached. For more information about this parameter, see UsageLimit.

modifyUsageLimit_usageLimitId :: Lens' ModifyUsageLimit Text

The identifier of the usage limit to modify.

Destructuring the Response

data UsageLimit

Describes a usage limit object for a cluster.

See: newUsageLimit smart constructor.

Constructors

UsageLimit'

Fields

amount :: Maybe Integer

The limit amount. If time-based, this amount is in minutes. If data-based, this amount is in terabytes (TB).

limitType :: Maybe UsageLimitLimitType

The type of limit. Depending on the feature type, this can be based on a time duration or data size.

usageLimitId :: Maybe Text

The identifier of the usage limit.

period :: Maybe UsageLimitPeriod

The time period that the amount applies to. A weekly period begins on Sunday. The default is monthly.

clusterIdentifier :: Maybe Text

The identifier of the cluster with a usage limit.

breachAction :: Maybe UsageLimitBreachAction

The action that Amazon Redshift takes when the limit is reached. Possible values are:

log - To log an event in a system table. The default is log.

emit-metric - To emit CloudWatch metrics.

disable - To disable the feature until the next usage period begins.

featureType :: Maybe UsageLimitFeatureType

The Amazon Redshift feature to which the limit applies.

tags :: Maybe [Tag]

A list of tag instances.

newUsageLimit :: UsageLimit

Create a value of UsageLimit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:amount:UsageLimit', usageLimit_amount - The limit amount. If time-based, this amount is in minutes. If data-based, this amount is in terabytes (TB).

$sel:limitType:UsageLimit', usageLimit_limitType - The type of limit. Depending on the feature type, this can be based on a time duration or data size.

$sel:usageLimitId:UsageLimit', usageLimit_usageLimitId - The identifier of the usage limit.

$sel:period:UsageLimit', usageLimit_period - The time period that the amount applies to. A weekly period begins on Sunday. The default is monthly.

$sel:clusterIdentifier:UsageLimit', usageLimit_clusterIdentifier - The identifier of the cluster with a usage limit.

$sel:breachAction:UsageLimit', usageLimit_breachAction - The action that Amazon Redshift takes when the limit is reached. Possible values are:

log - To log an event in a system table. The default is log.

emit-metric - To emit CloudWatch metrics.

disable - To disable the feature until the next usage period begins.

$sel:featureType:UsageLimit', usageLimit_featureType - The Amazon Redshift feature to which the limit applies.

$sel:tags:UsageLimit', usageLimit_tags - A list of tag instances.

Response Lenses

usageLimit_amount :: Lens' UsageLimit (Maybe Integer)

The limit amount. If time-based, this amount is in minutes. If data-based, this amount is in terabytes (TB).

usageLimit_limitType :: Lens' UsageLimit (Maybe UsageLimitLimitType)

The type of limit. Depending on the feature type, this can be based on a time duration or data size.

usageLimit_usageLimitId :: Lens' UsageLimit (Maybe Text)

The identifier of the usage limit.

usageLimit_period :: Lens' UsageLimit (Maybe UsageLimitPeriod)

The time period that the amount applies to. A weekly period begins on Sunday. The default is monthly.

usageLimit_clusterIdentifier :: Lens' UsageLimit (Maybe Text)

The identifier of the cluster with a usage limit.

usageLimit_breachAction :: Lens' UsageLimit (Maybe UsageLimitBreachAction)

The action that Amazon Redshift takes when the limit is reached. Possible values are:

log - To log an event in a system table. The default is log.

emit-metric - To emit CloudWatch metrics.

disable - To disable the feature until the next usage period begins.

usageLimit_featureType :: Lens' UsageLimit (Maybe UsageLimitFeatureType)

The Amazon Redshift feature to which the limit applies.

usageLimit_tags :: Lens' UsageLimit (Maybe [Tag])

A list of tag instances.