Network.AWS.Redshift.ResizeCluster

Description

Changes the size of the cluster. You can change the cluster's type, or change the number or type of nodes. The default behavior is to use the elastic resize method. With an elastic resize, your cluster is available for read and write operations more quickly than with the classic resize method.

Elastic resize operations have the following restrictions:

You can only resize clusters of the following types:

dc1.large (if your cluster is in a VPC)

dc1.8xlarge (if your cluster is in a VPC)

dc2.large

dc2.8xlarge

ds2.xlarge

ds2.8xlarge

ra3.xlplus

ra3.4xlarge

ra3.16xlarge

The type of nodes that you add must match the node type for the cluster.

Creating a Request

data ResizeCluster

Describes a resize cluster operation. For example, a scheduled action to run the ResizeCluster API operation.

See: newResizeCluster smart constructor.

Constructors

ResizeCluster'

Fields

numberOfNodes :: Maybe Int

The new number of nodes for the cluster. If not specified, the cluster's current number of nodes is used.

classic :: Maybe Bool

A boolean value indicating whether the resize operation is using the classic resize process. If you don't provide this parameter or set the value to false, the resize type is elastic.

clusterType :: Maybe Text

The new cluster type for the specified cluster.

nodeType :: Maybe Text

The new node type for the nodes you are adding. If not specified, the cluster's current node type is used.

clusterIdentifier :: Text

The unique identifier for the cluster to resize.

newResizeCluster

Arguments

:: Text
$sel:clusterIdentifier:ResizeCluster'

-> ResizeCluster

Create a value of ResizeCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:numberOfNodes:ResizeCluster', resizeCluster_numberOfNodes - The new number of nodes for the cluster. If not specified, the cluster's current number of nodes is used.

$sel:classic:ResizeCluster', resizeCluster_classic - A boolean value indicating whether the resize operation is using the classic resize process. If you don't provide this parameter or set the value to false, the resize type is elastic.

$sel:clusterType:ResizeCluster', resizeCluster_clusterType - The new cluster type for the specified cluster.

$sel:nodeType:ResizeCluster', resizeCluster_nodeType - The new node type for the nodes you are adding. If not specified, the cluster's current node type is used.

$sel:clusterIdentifier:ResizeCluster', resizeCluster_clusterIdentifier - The unique identifier for the cluster to resize.

Request Lenses

resizeCluster_numberOfNodes :: Lens' ResizeCluster (Maybe Int)

The new number of nodes for the cluster. If not specified, the cluster's current number of nodes is used.

resizeCluster_classic :: Lens' ResizeCluster (Maybe Bool)

A boolean value indicating whether the resize operation is using the classic resize process. If you don't provide this parameter or set the value to false, the resize type is elastic.

resizeCluster_clusterType :: Lens' ResizeCluster (Maybe Text)

The new cluster type for the specified cluster.

resizeCluster_nodeType :: Lens' ResizeCluster (Maybe Text)

The new node type for the nodes you are adding. If not specified, the cluster's current node type is used.

resizeCluster_clusterIdentifier :: Lens' ResizeCluster Text

The unique identifier for the cluster to resize.

Destructuring the Response

data ResizeClusterResponse

See: newResizeClusterResponse smart constructor.

Constructors

ResizeClusterResponse'

Fields

cluster :: Maybe Cluster

httpStatus :: Int

The response's http status code.

newResizeClusterResponse

Arguments

:: Int
$sel:httpStatus:ResizeClusterResponse'

-> ResizeClusterResponse

Create a value of ResizeClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:ResizeClusterResponse', resizeClusterResponse_cluster - Undocumented member.

$sel:httpStatus:ResizeClusterResponse', resizeClusterResponse_httpStatus - The response's http status code.

Response Lenses

resizeClusterResponse_cluster :: Lens' ResizeClusterResponse (Maybe Cluster)

Undocumented member.

resizeClusterResponse_httpStatus :: Lens' ResizeClusterResponse Int

The response's http status code.