Network.AWS.Redshift.RestoreFromClusterSnapshot

Description

Creates a new cluster from a snapshot. By default, Amazon Redshift creates the resulting cluster with the same configuration as the original cluster from which the snapshot was created, except that the new cluster is created with the default cluster security and parameter groups. After Amazon Redshift creates the cluster, you can use the ModifyCluster API to associate a different security group and different parameter group with the restored cluster. If you are using a DS node type, you can also choose to change to another DS node type of the same size during restore.

If you restore a cluster into a VPC, you must provide a cluster subnet group where you want the cluster restored.

For more information about working with snapshots, go to Amazon Redshift Snapshots in the Amazon Redshift Cluster Management Guide.

Creating a Request

data RestoreFromClusterSnapshot

See: newRestoreFromClusterSnapshot smart constructor.

Constructors

RestoreFromClusterSnapshot'

Fields

manualSnapshotRetentionPeriod :: Maybe Int

The default number of days to retain a manual snapshot. If the value is -1, the snapshot is retained indefinitely. This setting doesn't change the retention period of existing snapshots.

The value must be either -1 or an integer between 1 and 3,653.

enhancedVpcRouting :: Maybe Bool

An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

additionalInfo :: Maybe Text

Reserved.

snapshotScheduleIdentifier :: Maybe Text

A unique identifier for the snapshot schedule.

publiclyAccessible :: Maybe Bool

If true, the cluster can be accessed from a public network.

snapshotClusterIdentifier :: Maybe Text

The name of the cluster the source snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

maintenanceTrackName :: Maybe Text

The name of the maintenance track for the restored cluster. When you take a snapshot, the snapshot inherits the MaintenanceTrack value from the cluster. The snapshot might be on a different track than the cluster that was the source for the snapshot. For example, suppose that you take a snapshot of a cluster that is on the current track and then change the cluster to be on the trailing track. In this case, the snapshot and the source cluster are on different tracks.

hsmConfigurationIdentifier :: Maybe Text

Specifies the name of the HSM configuration that contains the information the Amazon Redshift cluster can use to retrieve and store keys in an HSM.

aquaConfigurationStatus :: Maybe AquaConfigurationStatus

The value represents how the cluster is configured to use AQUA (Advanced Query Accelerator) after the cluster is restored. Possible values include the following.

enabled - Use AQUA if it is available for the current Amazon Web Services Region and Amazon Redshift node type.

disabled - Don't use AQUA.

auto - Amazon Redshift determines whether to use AQUA.

clusterSecurityGroups :: Maybe [Text]

A list of security groups to be associated with this cluster.

Default: The default cluster security group for Amazon Redshift.

Cluster security groups only apply to clusters outside of VPCs.

automatedSnapshotRetentionPeriod :: Maybe Int

The number of days that automated snapshots are retained. If the value is 0, automated snapshots are disabled. Even if automated snapshots are disabled, you can still create manual snapshots when you want with CreateClusterSnapshot.

You can't disable automated snapshots for RA3 node types. Set the automated retention period from 1-35 days.

Default: The value selected for the cluster from which the snapshot was taken.

Constraints: Must be a value from 0 to 35.

clusterSubnetGroupName :: Maybe Text

The name of the subnet group where you want to cluster restored.

A snapshot of cluster in VPC can be restored only in VPC. Therefore, you must provide subnet group name where you want the cluster restored.

hsmClientCertificateIdentifier :: Maybe Text

Specifies the name of the HSM client certificate the Amazon Redshift cluster uses to retrieve the data encryption keys stored in an HSM.

numberOfNodes :: Maybe Int

The number of nodes specified when provisioning the restored cluster.

elasticIp :: Maybe Text

The elastic IP (EIP) address for the cluster.

preferredMaintenanceWindow :: Maybe Text

The weekly time range (in UTC) during which automated cluster maintenance can occur.

Format: ddd:hh24:mi-ddd:hh24:mi

Default: The value selected for the cluster from which the snapshot was taken. For more information about the time blocks for each region, see Maintenance Windows in Amazon Redshift Cluster Management Guide.

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Minimum 30-minute window.

kmsKeyId :: Maybe Text

The Key Management Service (KMS) key ID of the encryption key that you want to use to encrypt data in the cluster that you restore from a shared snapshot.

availabilityZone :: Maybe Text

The Amazon EC2 Availability Zone in which to restore the cluster.

Default: A random, system-chosen Availability Zone.

Example: us-east-2a

vpcSecurityGroupIds :: Maybe [Text]

A list of Virtual Private Cloud (VPC) security groups to be associated with the cluster.

Default: The default VPC security group is associated with the cluster.

VPC security groups only apply to clusters in VPCs.

iamRoles :: Maybe [Text]

A list of Identity and Access Management (IAM) roles that can be used by the cluster to access other Amazon Web Services services. You must supply the IAM roles in their Amazon Resource Name (ARN) format. You can supply up to 10 IAM roles in a single request.

A cluster can have up to 10 IAM roles associated at any time.

availabilityZoneRelocation :: Maybe Bool

The option to enable relocation for an Amazon Redshift cluster between Availability Zones after the cluster is restored.

ownerAccount :: Maybe Text

The Amazon Web Services account used to create or copy the snapshot. Required if you are restoring a snapshot you do not own, optional if you own the snapshot.

nodeType :: Maybe Text

The node type that the restored cluster will be provisioned with.

Default: The node type of the cluster from which the snapshot was taken. You can modify this if you are using any DS node type. In that case, you can choose to restore into another DS node type of the same size. For example, you can restore ds1.8xlarge into ds2.8xlarge, or ds1.xlarge into ds2.xlarge. If you have a DC instance type, you must restore into that same instance type and size. In other words, you can only restore a dc1.large instance type into another dc1.large instance type or dc2.large instance type. You can't restore dc1.8xlarge to dc2.8xlarge. First restore to a dc1.8xlarge cluster, then resize to a dc2.8large cluster. For more information about node types, see About Clusters and Nodes in the Amazon Redshift Cluster Management Guide.

allowVersionUpgrade :: Maybe Bool

If true, major version upgrades can be applied during the maintenance window to the Amazon Redshift engine that is running on the cluster.

Default: true

clusterParameterGroupName :: Maybe Text

The name of the parameter group to be associated with this cluster.

Default: The default Amazon Redshift cluster parameter group. For information about the default parameter group, go to Working with Amazon Redshift Parameter Groups.

Constraints:

Must be 1 to 255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

port :: Maybe Int

The port number on which the cluster accepts connections.

Default: The same port as the original cluster.

Constraints: Must be between 1115 and 65535.

clusterIdentifier :: Text

The identifier of the cluster that will be created from restoring the snapshot.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

Alphabetic characters must be lowercase.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Must be unique for all clusters within an Amazon Web Services account.

snapshotIdentifier :: Text

The name of the snapshot from which to create the new cluster. This parameter isn't case sensitive.

Example: my-snapshot-id

newRestoreFromClusterSnapshot

Arguments

:: Text
$sel:clusterIdentifier:RestoreFromClusterSnapshot'

-> Text
$sel:snapshotIdentifier:RestoreFromClusterSnapshot'

-> RestoreFromClusterSnapshot

Create a value of RestoreFromClusterSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:manualSnapshotRetentionPeriod:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_manualSnapshotRetentionPeriod - The default number of days to retain a manual snapshot. If the value is -1, the snapshot is retained indefinitely. This setting doesn't change the retention period of existing snapshots.

The value must be either -1 or an integer between 1 and 3,653.

$sel:enhancedVpcRouting:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_enhancedVpcRouting - An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

$sel:additionalInfo:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_additionalInfo - Reserved.

$sel:snapshotScheduleIdentifier:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_snapshotScheduleIdentifier - A unique identifier for the snapshot schedule.

$sel:publiclyAccessible:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_publiclyAccessible - If true, the cluster can be accessed from a public network.

$sel:snapshotClusterIdentifier:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_snapshotClusterIdentifier - The name of the cluster the source snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

$sel:maintenanceTrackName:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_maintenanceTrackName - The name of the maintenance track for the restored cluster. When you take a snapshot, the snapshot inherits the MaintenanceTrack value from the cluster. The snapshot might be on a different track than the cluster that was the source for the snapshot. For example, suppose that you take a snapshot of a cluster that is on the current track and then change the cluster to be on the trailing track. In this case, the snapshot and the source cluster are on different tracks.

$sel:hsmConfigurationIdentifier:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_hsmConfigurationIdentifier - Specifies the name of the HSM configuration that contains the information the Amazon Redshift cluster can use to retrieve and store keys in an HSM.

$sel:aquaConfigurationStatus:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_aquaConfigurationStatus - The value represents how the cluster is configured to use AQUA (Advanced Query Accelerator) after the cluster is restored. Possible values include the following.

enabled - Use AQUA if it is available for the current Amazon Web Services Region and Amazon Redshift node type.

disabled - Don't use AQUA.

auto - Amazon Redshift determines whether to use AQUA.

$sel:clusterSecurityGroups:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_clusterSecurityGroups - A list of security groups to be associated with this cluster.

Default: The default cluster security group for Amazon Redshift.

Cluster security groups only apply to clusters outside of VPCs.

$sel:automatedSnapshotRetentionPeriod:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_automatedSnapshotRetentionPeriod - The number of days that automated snapshots are retained. If the value is 0, automated snapshots are disabled. Even if automated snapshots are disabled, you can still create manual snapshots when you want with CreateClusterSnapshot.

You can't disable automated snapshots for RA3 node types. Set the automated retention period from 1-35 days.

Default: The value selected for the cluster from which the snapshot was taken.

Constraints: Must be a value from 0 to 35.

$sel:clusterSubnetGroupName:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_clusterSubnetGroupName - The name of the subnet group where you want to cluster restored.

A snapshot of cluster in VPC can be restored only in VPC. Therefore, you must provide subnet group name where you want the cluster restored.

$sel:hsmClientCertificateIdentifier:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_hsmClientCertificateIdentifier - Specifies the name of the HSM client certificate the Amazon Redshift cluster uses to retrieve the data encryption keys stored in an HSM.

$sel:numberOfNodes:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_numberOfNodes - The number of nodes specified when provisioning the restored cluster.

$sel:elasticIp:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_elasticIp - The elastic IP (EIP) address for the cluster.

$sel:preferredMaintenanceWindow:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_preferredMaintenanceWindow - The weekly time range (in UTC) during which automated cluster maintenance can occur.

Format: ddd:hh24:mi-ddd:hh24:mi

Default: The value selected for the cluster from which the snapshot was taken. For more information about the time blocks for each region, see Maintenance Windows in Amazon Redshift Cluster Management Guide.

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Minimum 30-minute window.

$sel:kmsKeyId:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_kmsKeyId - The Key Management Service (KMS) key ID of the encryption key that you want to use to encrypt data in the cluster that you restore from a shared snapshot.

$sel:availabilityZone:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_availabilityZone - The Amazon EC2 Availability Zone in which to restore the cluster.

Default: A random, system-chosen Availability Zone.

Example: us-east-2a

$sel:vpcSecurityGroupIds:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_vpcSecurityGroupIds - A list of Virtual Private Cloud (VPC) security groups to be associated with the cluster.

Default: The default VPC security group is associated with the cluster.

VPC security groups only apply to clusters in VPCs.

$sel:iamRoles:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_iamRoles - A list of Identity and Access Management (IAM) roles that can be used by the cluster to access other Amazon Web Services services. You must supply the IAM roles in their Amazon Resource Name (ARN) format. You can supply up to 10 IAM roles in a single request.

A cluster can have up to 10 IAM roles associated at any time.

$sel:availabilityZoneRelocation:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_availabilityZoneRelocation - The option to enable relocation for an Amazon Redshift cluster between Availability Zones after the cluster is restored.

$sel:ownerAccount:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_ownerAccount - The Amazon Web Services account used to create or copy the snapshot. Required if you are restoring a snapshot you do not own, optional if you own the snapshot.

$sel:nodeType:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_nodeType - The node type that the restored cluster will be provisioned with.

Default: The node type of the cluster from which the snapshot was taken. You can modify this if you are using any DS node type. In that case, you can choose to restore into another DS node type of the same size. For example, you can restore ds1.8xlarge into ds2.8xlarge, or ds1.xlarge into ds2.xlarge. If you have a DC instance type, you must restore into that same instance type and size. In other words, you can only restore a dc1.large instance type into another dc1.large instance type or dc2.large instance type. You can't restore dc1.8xlarge to dc2.8xlarge. First restore to a dc1.8xlarge cluster, then resize to a dc2.8large cluster. For more information about node types, see About Clusters and Nodes in the Amazon Redshift Cluster Management Guide.

$sel:allowVersionUpgrade:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_allowVersionUpgrade - If true, major version upgrades can be applied during the maintenance window to the Amazon Redshift engine that is running on the cluster.

Default: true

$sel:clusterParameterGroupName:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_clusterParameterGroupName - The name of the parameter group to be associated with this cluster.

Default: The default Amazon Redshift cluster parameter group. For information about the default parameter group, go to Working with Amazon Redshift Parameter Groups.

Constraints:

Must be 1 to 255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

$sel:port:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_port - The port number on which the cluster accepts connections.

Default: The same port as the original cluster.

Constraints: Must be between 1115 and 65535.

$sel:clusterIdentifier:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_clusterIdentifier - The identifier of the cluster that will be created from restoring the snapshot.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

Alphabetic characters must be lowercase.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Must be unique for all clusters within an Amazon Web Services account.

$sel:snapshotIdentifier:RestoreFromClusterSnapshot', restoreFromClusterSnapshot_snapshotIdentifier - The name of the snapshot from which to create the new cluster. This parameter isn't case sensitive.

Example: my-snapshot-id

Request Lenses

restoreFromClusterSnapshot_manualSnapshotRetentionPeriod :: Lens' RestoreFromClusterSnapshot (Maybe Int)

The default number of days to retain a manual snapshot. If the value is -1, the snapshot is retained indefinitely. This setting doesn't change the retention period of existing snapshots.

The value must be either -1 or an integer between 1 and 3,653.

restoreFromClusterSnapshot_enhancedVpcRouting :: Lens' RestoreFromClusterSnapshot (Maybe Bool)

An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

restoreFromClusterSnapshot_additionalInfo :: Lens' RestoreFromClusterSnapshot (Maybe Text)

Reserved.

restoreFromClusterSnapshot_snapshotScheduleIdentifier :: Lens' RestoreFromClusterSnapshot (Maybe Text)

A unique identifier for the snapshot schedule.

restoreFromClusterSnapshot_publiclyAccessible :: Lens' RestoreFromClusterSnapshot (Maybe Bool)

If true, the cluster can be accessed from a public network.

restoreFromClusterSnapshot_snapshotClusterIdentifier :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The name of the cluster the source snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

restoreFromClusterSnapshot_maintenanceTrackName :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The name of the maintenance track for the restored cluster. When you take a snapshot, the snapshot inherits the MaintenanceTrack value from the cluster. The snapshot might be on a different track than the cluster that was the source for the snapshot. For example, suppose that you take a snapshot of a cluster that is on the current track and then change the cluster to be on the trailing track. In this case, the snapshot and the source cluster are on different tracks.

restoreFromClusterSnapshot_hsmConfigurationIdentifier :: Lens' RestoreFromClusterSnapshot (Maybe Text)

Specifies the name of the HSM configuration that contains the information the Amazon Redshift cluster can use to retrieve and store keys in an HSM.

restoreFromClusterSnapshot_aquaConfigurationStatus :: Lens' RestoreFromClusterSnapshot (Maybe AquaConfigurationStatus)

The value represents how the cluster is configured to use AQUA (Advanced Query Accelerator) after the cluster is restored. Possible values include the following.

enabled - Use AQUA if it is available for the current Amazon Web Services Region and Amazon Redshift node type.

disabled - Don't use AQUA.

auto - Amazon Redshift determines whether to use AQUA.

restoreFromClusterSnapshot_clusterSecurityGroups :: Lens' RestoreFromClusterSnapshot (Maybe [Text])

A list of security groups to be associated with this cluster.

Default: The default cluster security group for Amazon Redshift.

Cluster security groups only apply to clusters outside of VPCs.

restoreFromClusterSnapshot_automatedSnapshotRetentionPeriod :: Lens' RestoreFromClusterSnapshot (Maybe Int)

The number of days that automated snapshots are retained. If the value is 0, automated snapshots are disabled. Even if automated snapshots are disabled, you can still create manual snapshots when you want with CreateClusterSnapshot.

You can't disable automated snapshots for RA3 node types. Set the automated retention period from 1-35 days.

Default: The value selected for the cluster from which the snapshot was taken.

Constraints: Must be a value from 0 to 35.

restoreFromClusterSnapshot_clusterSubnetGroupName :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The name of the subnet group where you want to cluster restored.

A snapshot of cluster in VPC can be restored only in VPC. Therefore, you must provide subnet group name where you want the cluster restored.

restoreFromClusterSnapshot_hsmClientCertificateIdentifier :: Lens' RestoreFromClusterSnapshot (Maybe Text)

Specifies the name of the HSM client certificate the Amazon Redshift cluster uses to retrieve the data encryption keys stored in an HSM.

restoreFromClusterSnapshot_numberOfNodes :: Lens' RestoreFromClusterSnapshot (Maybe Int)

The number of nodes specified when provisioning the restored cluster.

restoreFromClusterSnapshot_elasticIp :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The elastic IP (EIP) address for the cluster.

restoreFromClusterSnapshot_preferredMaintenanceWindow :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The weekly time range (in UTC) during which automated cluster maintenance can occur.

Format: ddd:hh24:mi-ddd:hh24:mi

Default: The value selected for the cluster from which the snapshot was taken. For more information about the time blocks for each region, see Maintenance Windows in Amazon Redshift Cluster Management Guide.

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Minimum 30-minute window.

restoreFromClusterSnapshot_kmsKeyId :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The Key Management Service (KMS) key ID of the encryption key that you want to use to encrypt data in the cluster that you restore from a shared snapshot.

restoreFromClusterSnapshot_availabilityZone :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The Amazon EC2 Availability Zone in which to restore the cluster.

Default: A random, system-chosen Availability Zone.

Example: us-east-2a

restoreFromClusterSnapshot_vpcSecurityGroupIds :: Lens' RestoreFromClusterSnapshot (Maybe [Text])

A list of Virtual Private Cloud (VPC) security groups to be associated with the cluster.

Default: The default VPC security group is associated with the cluster.

VPC security groups only apply to clusters in VPCs.

restoreFromClusterSnapshot_iamRoles :: Lens' RestoreFromClusterSnapshot (Maybe [Text])

A list of Identity and Access Management (IAM) roles that can be used by the cluster to access other Amazon Web Services services. You must supply the IAM roles in their Amazon Resource Name (ARN) format. You can supply up to 10 IAM roles in a single request.

A cluster can have up to 10 IAM roles associated at any time.

restoreFromClusterSnapshot_availabilityZoneRelocation :: Lens' RestoreFromClusterSnapshot (Maybe Bool)

The option to enable relocation for an Amazon Redshift cluster between Availability Zones after the cluster is restored.

restoreFromClusterSnapshot_ownerAccount :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The Amazon Web Services account used to create or copy the snapshot. Required if you are restoring a snapshot you do not own, optional if you own the snapshot.

restoreFromClusterSnapshot_nodeType :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The node type that the restored cluster will be provisioned with.

Default: The node type of the cluster from which the snapshot was taken. You can modify this if you are using any DS node type. In that case, you can choose to restore into another DS node type of the same size. For example, you can restore ds1.8xlarge into ds2.8xlarge, or ds1.xlarge into ds2.xlarge. If you have a DC instance type, you must restore into that same instance type and size. In other words, you can only restore a dc1.large instance type into another dc1.large instance type or dc2.large instance type. You can't restore dc1.8xlarge to dc2.8xlarge. First restore to a dc1.8xlarge cluster, then resize to a dc2.8large cluster. For more information about node types, see About Clusters and Nodes in the Amazon Redshift Cluster Management Guide.

restoreFromClusterSnapshot_allowVersionUpgrade :: Lens' RestoreFromClusterSnapshot (Maybe Bool)

If true, major version upgrades can be applied during the maintenance window to the Amazon Redshift engine that is running on the cluster.

Default: true

restoreFromClusterSnapshot_clusterParameterGroupName :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The name of the parameter group to be associated with this cluster.

Default: The default Amazon Redshift cluster parameter group. For information about the default parameter group, go to Working with Amazon Redshift Parameter Groups.

Constraints:

Must be 1 to 255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

restoreFromClusterSnapshot_port :: Lens' RestoreFromClusterSnapshot (Maybe Int)

The port number on which the cluster accepts connections.

Default: The same port as the original cluster.

Constraints: Must be between 1115 and 65535.

restoreFromClusterSnapshot_clusterIdentifier :: Lens' RestoreFromClusterSnapshot Text

The identifier of the cluster that will be created from restoring the snapshot.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

Alphabetic characters must be lowercase.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Must be unique for all clusters within an Amazon Web Services account.

restoreFromClusterSnapshot_snapshotIdentifier :: Lens' RestoreFromClusterSnapshot Text

The name of the snapshot from which to create the new cluster. This parameter isn't case sensitive.

Example: my-snapshot-id

Destructuring the Response

data RestoreFromClusterSnapshotResponse

See: newRestoreFromClusterSnapshotResponse smart constructor.

Constructors

RestoreFromClusterSnapshotResponse'

Fields

cluster :: Maybe Cluster

httpStatus :: Int

The response's http status code.

newRestoreFromClusterSnapshotResponse

Arguments

:: Int
$sel:httpStatus:RestoreFromClusterSnapshotResponse'

-> RestoreFromClusterSnapshotResponse

Create a value of RestoreFromClusterSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:RestoreFromClusterSnapshotResponse', restoreFromClusterSnapshotResponse_cluster - Undocumented member.

$sel:httpStatus:RestoreFromClusterSnapshotResponse', restoreFromClusterSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

restoreFromClusterSnapshotResponse_cluster :: Lens' RestoreFromClusterSnapshotResponse (Maybe Cluster)

Undocumented member.

restoreFromClusterSnapshotResponse_httpStatus :: Lens' RestoreFromClusterSnapshotResponse Int

The response's http status code.