Network.AWS.Redshift.RestoreTableFromClusterSnapshot

Description

Creates a new table from a table in an Amazon Redshift cluster snapshot. You must create the new table within the Amazon Redshift cluster that the snapshot was taken from.

You cannot use RestoreTableFromClusterSnapshot to restore a table with the same name as an existing table in an Amazon Redshift cluster. That is, you cannot overwrite an existing table in a cluster with a restored table. If you want to replace your original table with a new, restored table, then rename or drop your original table before you call RestoreTableFromClusterSnapshot. When you have renamed your original table, then you can pass the original name of the table as the NewTableName parameter value in the call to RestoreTableFromClusterSnapshot. This way, you can replace the original table with the table created from the snapshot.

Creating a Request

data RestoreTableFromClusterSnapshot

See: newRestoreTableFromClusterSnapshot smart constructor.

Constructors

RestoreTableFromClusterSnapshot'

Fields

targetSchemaName :: Maybe Text

The name of the schema to restore the table to.

sourceSchemaName :: Maybe Text

The name of the source schema that contains the table to restore from. If you do not specify a SourceSchemaName value, the default is public.

targetDatabaseName :: Maybe Text

The name of the database to restore the table to.

enableCaseSensitiveIdentifier :: Maybe Bool

Indicates whether name identifiers for database, schema, and table are case sensitive. If true, the names are case sensitive. If false (default), the names are not case sensitive.

clusterIdentifier :: Text

The identifier of the Amazon Redshift cluster to restore the table to.

snapshotIdentifier :: Text

The identifier of the snapshot to restore the table from. This snapshot must have been created from the Amazon Redshift cluster specified by the ClusterIdentifier parameter.

sourceDatabaseName :: Text

The name of the source database that contains the table to restore from.

sourceTableName :: Text

The name of the source table to restore from.

newTableName' :: Text

The name of the table to create as a result of the current request.

newRestoreTableFromClusterSnapshot

Arguments

:: Text
$sel:clusterIdentifier:RestoreTableFromClusterSnapshot'

-> Text
$sel:snapshotIdentifier:RestoreTableFromClusterSnapshot'

-> Text
$sel:sourceDatabaseName:RestoreTableFromClusterSnapshot'

-> Text
$sel:sourceTableName:RestoreTableFromClusterSnapshot'

-> Text
$sel:newTableName':RestoreTableFromClusterSnapshot'

-> RestoreTableFromClusterSnapshot

Create a value of RestoreTableFromClusterSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetSchemaName:RestoreTableFromClusterSnapshot', restoreTableFromClusterSnapshot_targetSchemaName - The name of the schema to restore the table to.

$sel:sourceSchemaName:RestoreTableFromClusterSnapshot', restoreTableFromClusterSnapshot_sourceSchemaName - The name of the source schema that contains the table to restore from. If you do not specify a SourceSchemaName value, the default is public.

$sel:targetDatabaseName:RestoreTableFromClusterSnapshot', restoreTableFromClusterSnapshot_targetDatabaseName - The name of the database to restore the table to.

$sel:enableCaseSensitiveIdentifier:RestoreTableFromClusterSnapshot', restoreTableFromClusterSnapshot_enableCaseSensitiveIdentifier - Indicates whether name identifiers for database, schema, and table are case sensitive. If true, the names are case sensitive. If false (default), the names are not case sensitive.

$sel:clusterIdentifier:RestoreTableFromClusterSnapshot', restoreTableFromClusterSnapshot_clusterIdentifier - The identifier of the Amazon Redshift cluster to restore the table to.

$sel:snapshotIdentifier:RestoreTableFromClusterSnapshot', restoreTableFromClusterSnapshot_snapshotIdentifier - The identifier of the snapshot to restore the table from. This snapshot must have been created from the Amazon Redshift cluster specified by the ClusterIdentifier parameter.

$sel:sourceDatabaseName:RestoreTableFromClusterSnapshot', restoreTableFromClusterSnapshot_sourceDatabaseName - The name of the source database that contains the table to restore from.

$sel:sourceTableName:RestoreTableFromClusterSnapshot', restoreTableFromClusterSnapshot_sourceTableName - The name of the source table to restore from.

$sel:newTableName':RestoreTableFromClusterSnapshot', restoreTableFromClusterSnapshot_newTableName - The name of the table to create as a result of the current request.

Request Lenses

restoreTableFromClusterSnapshot_targetSchemaName :: Lens' RestoreTableFromClusterSnapshot (Maybe Text)

The name of the schema to restore the table to.

restoreTableFromClusterSnapshot_sourceSchemaName :: Lens' RestoreTableFromClusterSnapshot (Maybe Text)

The name of the source schema that contains the table to restore from. If you do not specify a SourceSchemaName value, the default is public.

restoreTableFromClusterSnapshot_targetDatabaseName :: Lens' RestoreTableFromClusterSnapshot (Maybe Text)

The name of the database to restore the table to.

restoreTableFromClusterSnapshot_enableCaseSensitiveIdentifier :: Lens' RestoreTableFromClusterSnapshot (Maybe Bool)

Indicates whether name identifiers for database, schema, and table are case sensitive. If true, the names are case sensitive. If false (default), the names are not case sensitive.

restoreTableFromClusterSnapshot_clusterIdentifier :: Lens' RestoreTableFromClusterSnapshot Text

The identifier of the Amazon Redshift cluster to restore the table to.

restoreTableFromClusterSnapshot_snapshotIdentifier :: Lens' RestoreTableFromClusterSnapshot Text

The identifier of the snapshot to restore the table from. This snapshot must have been created from the Amazon Redshift cluster specified by the ClusterIdentifier parameter.

restoreTableFromClusterSnapshot_sourceDatabaseName :: Lens' RestoreTableFromClusterSnapshot Text

The name of the source database that contains the table to restore from.

restoreTableFromClusterSnapshot_sourceTableName :: Lens' RestoreTableFromClusterSnapshot Text

The name of the source table to restore from.

restoreTableFromClusterSnapshot_newTableName :: Lens' RestoreTableFromClusterSnapshot Text

The name of the table to create as a result of the current request.

Destructuring the Response

data RestoreTableFromClusterSnapshotResponse

See: newRestoreTableFromClusterSnapshotResponse smart constructor.

Constructors

RestoreTableFromClusterSnapshotResponse'

Fields

tableRestoreStatus :: Maybe TableRestoreStatus

httpStatus :: Int

The response's http status code.

newRestoreTableFromClusterSnapshotResponse

Arguments

:: Int
$sel:httpStatus:RestoreTableFromClusterSnapshotResponse'

-> RestoreTableFromClusterSnapshotResponse

Create a value of RestoreTableFromClusterSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tableRestoreStatus:RestoreTableFromClusterSnapshotResponse', restoreTableFromClusterSnapshotResponse_tableRestoreStatus - Undocumented member.

$sel:httpStatus:RestoreTableFromClusterSnapshotResponse', restoreTableFromClusterSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

restoreTableFromClusterSnapshotResponse_tableRestoreStatus :: Lens' RestoreTableFromClusterSnapshotResponse (Maybe TableRestoreStatus)

Undocumented member.

restoreTableFromClusterSnapshotResponse_httpStatus :: Lens' RestoreTableFromClusterSnapshotResponse Int

The response's http status code.