Network.AWS.Redshift.RevokeSnapshotAccess

Description

Removes the ability of the specified Amazon Web Services account to restore the specified snapshot. If the account is currently restoring the snapshot, the restore will run to completion.

For more information about working with snapshots, go to Amazon Redshift Snapshots in the Amazon Redshift Cluster Management Guide.

Creating a Request

data RevokeSnapshotAccess

See: newRevokeSnapshotAccess smart constructor.

Constructors

RevokeSnapshotAccess'

Fields

snapshotClusterIdentifier :: Maybe Text

The identifier of the cluster the snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

snapshotIdentifier :: Text

The identifier of the snapshot that the account can no longer access.

accountWithRestoreAccess :: Text

The identifier of the Amazon Web Services account that can no longer restore the specified snapshot.

newRevokeSnapshotAccess

Arguments

:: Text
$sel:snapshotIdentifier:RevokeSnapshotAccess'

-> Text
$sel:accountWithRestoreAccess:RevokeSnapshotAccess'

-> RevokeSnapshotAccess

Create a value of RevokeSnapshotAccess with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshotClusterIdentifier:RevokeSnapshotAccess', revokeSnapshotAccess_snapshotClusterIdentifier - The identifier of the cluster the snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

$sel:snapshotIdentifier:RevokeSnapshotAccess', revokeSnapshotAccess_snapshotIdentifier - The identifier of the snapshot that the account can no longer access.

$sel:accountWithRestoreAccess:RevokeSnapshotAccess', revokeSnapshotAccess_accountWithRestoreAccess - The identifier of the Amazon Web Services account that can no longer restore the specified snapshot.

Request Lenses

revokeSnapshotAccess_snapshotClusterIdentifier :: Lens' RevokeSnapshotAccess (Maybe Text)

The identifier of the cluster the snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

revokeSnapshotAccess_snapshotIdentifier :: Lens' RevokeSnapshotAccess Text

The identifier of the snapshot that the account can no longer access.

revokeSnapshotAccess_accountWithRestoreAccess :: Lens' RevokeSnapshotAccess Text

The identifier of the Amazon Web Services account that can no longer restore the specified snapshot.

Destructuring the Response

data RevokeSnapshotAccessResponse

See: newRevokeSnapshotAccessResponse smart constructor.

Constructors

RevokeSnapshotAccessResponse'

Fields

snapshot :: Maybe Snapshot

httpStatus :: Int

The response's http status code.

newRevokeSnapshotAccessResponse

Arguments

:: Int
$sel:httpStatus:RevokeSnapshotAccessResponse'

-> RevokeSnapshotAccessResponse

Create a value of RevokeSnapshotAccessResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshot:RevokeSnapshotAccessResponse', revokeSnapshotAccessResponse_snapshot - Undocumented member.

$sel:httpStatus:RevokeSnapshotAccessResponse', revokeSnapshotAccessResponse_httpStatus - The response's http status code.

Response Lenses

revokeSnapshotAccessResponse_snapshot :: Lens' RevokeSnapshotAccessResponse (Maybe Snapshot)

Undocumented member.

revokeSnapshotAccessResponse_httpStatus :: Lens' RevokeSnapshotAccessResponse Int

The response's http status code.