Network.AWS.Redshift.Types.ClusterSecurityGroup

Description

Documentation

data ClusterSecurityGroup

Describes a security group.

See: newClusterSecurityGroup smart constructor.

Constructors

ClusterSecurityGroup'

Fields

iPRanges :: Maybe [IPRange]

A list of IP ranges (CIDR blocks) that are permitted to access clusters associated with this cluster security group.

clusterSecurityGroupName :: Maybe Text

The name of the cluster security group to which the operation was applied.

tags :: Maybe [Tag]

The list of tags for the cluster security group.

eC2SecurityGroups :: Maybe [EC2SecurityGroup]

A list of EC2 security groups that are permitted to access clusters associated with this cluster security group.

description :: Maybe Text

A description of the security group.

newClusterSecurityGroup :: ClusterSecurityGroup

Create a value of ClusterSecurityGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iPRanges:ClusterSecurityGroup', clusterSecurityGroup_iPRanges - A list of IP ranges (CIDR blocks) that are permitted to access clusters associated with this cluster security group.

$sel:clusterSecurityGroupName:ClusterSecurityGroup', clusterSecurityGroup_clusterSecurityGroupName - The name of the cluster security group to which the operation was applied.

$sel:tags:ClusterSecurityGroup', clusterSecurityGroup_tags - The list of tags for the cluster security group.

$sel:eC2SecurityGroups:ClusterSecurityGroup', clusterSecurityGroup_eC2SecurityGroups - A list of EC2 security groups that are permitted to access clusters associated with this cluster security group.

$sel:description:ClusterSecurityGroup', clusterSecurityGroup_description - A description of the security group.

clusterSecurityGroup_iPRanges :: Lens' ClusterSecurityGroup (Maybe [IPRange])

A list of IP ranges (CIDR blocks) that are permitted to access clusters associated with this cluster security group.

clusterSecurityGroup_clusterSecurityGroupName :: Lens' ClusterSecurityGroup (Maybe Text)

The name of the cluster security group to which the operation was applied.

clusterSecurityGroup_tags :: Lens' ClusterSecurityGroup (Maybe [Tag])

The list of tags for the cluster security group.

clusterSecurityGroup_eC2SecurityGroups :: Lens' ClusterSecurityGroup (Maybe [EC2SecurityGroup])

A list of EC2 security groups that are permitted to access clusters associated with this cluster security group.

clusterSecurityGroup_description :: Lens' ClusterSecurityGroup (Maybe Text)

A description of the security group.