Network.AWS.Redshift.Types.ClusterSnapshotCopyStatus

Description

Documentation

data ClusterSnapshotCopyStatus

Returns the destination region and retention period that are configured for cross-region snapshot copy.

See: newClusterSnapshotCopyStatus smart constructor.

Constructors

ClusterSnapshotCopyStatus'

Fields

destinationRegion :: Maybe Text

The destination region that snapshots are automatically copied to when cross-region snapshot copy is enabled.

snapshotCopyGrantName :: Maybe Text

The name of the snapshot copy grant.

manualSnapshotRetentionPeriod :: Maybe Int

The number of days that automated snapshots are retained in the destination region after they are copied from a source region. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

retentionPeriod :: Maybe Integer

The number of days that automated snapshots are retained in the destination region after they are copied from a source region.

newClusterSnapshotCopyStatus :: ClusterSnapshotCopyStatus

Create a value of ClusterSnapshotCopyStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationRegion:ClusterSnapshotCopyStatus', clusterSnapshotCopyStatus_destinationRegion - The destination region that snapshots are automatically copied to when cross-region snapshot copy is enabled.

$sel:snapshotCopyGrantName:ClusterSnapshotCopyStatus', clusterSnapshotCopyStatus_snapshotCopyGrantName - The name of the snapshot copy grant.

$sel:manualSnapshotRetentionPeriod:ClusterSnapshotCopyStatus', clusterSnapshotCopyStatus_manualSnapshotRetentionPeriod - The number of days that automated snapshots are retained in the destination region after they are copied from a source region. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

$sel:retentionPeriod:ClusterSnapshotCopyStatus', clusterSnapshotCopyStatus_retentionPeriod - The number of days that automated snapshots are retained in the destination region after they are copied from a source region.

clusterSnapshotCopyStatus_destinationRegion :: Lens' ClusterSnapshotCopyStatus (Maybe Text)

The destination region that snapshots are automatically copied to when cross-region snapshot copy is enabled.

clusterSnapshotCopyStatus_snapshotCopyGrantName :: Lens' ClusterSnapshotCopyStatus (Maybe Text)

The name of the snapshot copy grant.

clusterSnapshotCopyStatus_manualSnapshotRetentionPeriod :: Lens' ClusterSnapshotCopyStatus (Maybe Int)

The number of days that automated snapshots are retained in the destination region after they are copied from a source region. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

clusterSnapshotCopyStatus_retentionPeriod :: Lens' ClusterSnapshotCopyStatus (Maybe Integer)

The number of days that automated snapshots are retained in the destination region after they are copied from a source region.