Network.AWS.Redshift.Types.ClusterSubnetGroup

Description

Documentation

data ClusterSubnetGroup

Describes a subnet group.

See: newClusterSubnetGroup smart constructor.

Constructors

ClusterSubnetGroup'

Fields

clusterSubnetGroupName :: Maybe Text

The name of the cluster subnet group.

subnetGroupStatus :: Maybe Text

The status of the cluster subnet group. Possible values are Complete, Incomplete and Invalid.

tags :: Maybe [Tag]

The list of tags for the cluster subnet group.

description :: Maybe Text

The description of the cluster subnet group.

vpcId :: Maybe Text

The VPC ID of the cluster subnet group.

subnets :: Maybe [Subnet]

A list of the VPC Subnet elements.

newClusterSubnetGroup :: ClusterSubnetGroup

Create a value of ClusterSubnetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterSubnetGroupName:ClusterSubnetGroup', clusterSubnetGroup_clusterSubnetGroupName - The name of the cluster subnet group.

$sel:subnetGroupStatus:ClusterSubnetGroup', clusterSubnetGroup_subnetGroupStatus - The status of the cluster subnet group. Possible values are Complete, Incomplete and Invalid.

$sel:tags:ClusterSubnetGroup', clusterSubnetGroup_tags - The list of tags for the cluster subnet group.

$sel:description:ClusterSubnetGroup', clusterSubnetGroup_description - The description of the cluster subnet group.

$sel:vpcId:ClusterSubnetGroup', clusterSubnetGroup_vpcId - The VPC ID of the cluster subnet group.

$sel:subnets:ClusterSubnetGroup', clusterSubnetGroup_subnets - A list of the VPC Subnet elements.

clusterSubnetGroup_clusterSubnetGroupName :: Lens' ClusterSubnetGroup (Maybe Text)

The name of the cluster subnet group.

clusterSubnetGroup_subnetGroupStatus :: Lens' ClusterSubnetGroup (Maybe Text)

The status of the cluster subnet group. Possible values are Complete, Incomplete and Invalid.

clusterSubnetGroup_tags :: Lens' ClusterSubnetGroup (Maybe [Tag])

The list of tags for the cluster subnet group.

clusterSubnetGroup_description :: Lens' ClusterSubnetGroup (Maybe Text)

The description of the cluster subnet group.

clusterSubnetGroup_vpcId :: Lens' ClusterSubnetGroup (Maybe Text)

The VPC ID of the cluster subnet group.

clusterSubnetGroup_subnets :: Lens' ClusterSubnetGroup (Maybe [Subnet])

A list of the VPC Subnet elements.