Network.AWS.Redshift.Types.DataShareAssociation

Description

Documentation

data DataShareAssociation

The association of a datashare from a producer account with a data consumer.

See: newDataShareAssociation smart constructor.

Constructors

DataShareAssociation'

Fields

createdDate :: Maybe ISO8601

The creation date of the datashare that is associated.

status :: Maybe DataShareStatus

The status of the datashare that is associated.

statusChangeDate :: Maybe ISO8601

The status change data of the datashare that is associated.

consumerIdentifier :: Maybe Text

The name of the consumer accounts that have an association with a producer datashare.

newDataShareAssociation :: DataShareAssociation

Create a value of DataShareAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdDate:DataShareAssociation', dataShareAssociation_createdDate - The creation date of the datashare that is associated.

$sel:status:DataShareAssociation', dataShareAssociation_status - The status of the datashare that is associated.

$sel:statusChangeDate:DataShareAssociation', dataShareAssociation_statusChangeDate - The status change data of the datashare that is associated.

$sel:consumerIdentifier:DataShareAssociation', dataShareAssociation_consumerIdentifier - The name of the consumer accounts that have an association with a producer datashare.

dataShareAssociation_createdDate :: Lens' DataShareAssociation (Maybe UTCTime)

The creation date of the datashare that is associated.

dataShareAssociation_status :: Lens' DataShareAssociation (Maybe DataShareStatus)

The status of the datashare that is associated.

dataShareAssociation_statusChangeDate :: Lens' DataShareAssociation (Maybe UTCTime)

The status change data of the datashare that is associated.

dataShareAssociation_consumerIdentifier :: Lens' DataShareAssociation (Maybe Text)

The name of the consumer accounts that have an association with a producer datashare.