Network.AWS.Redshift.Types.DataTransferProgress

Description

Documentation

data DataTransferProgress

Describes the status of a cluster while it is in the process of resizing with an incremental resize.

See: newDataTransferProgress smart constructor.

Constructors

DataTransferProgress'

Fields

status :: Maybe Text

Describes the status of the cluster. While the transfer is in progress the status is transferringdata.

estimatedTimeToCompletionInSeconds :: Maybe Integer

Describes the estimated number of seconds remaining to complete the transfer.

dataTransferredInMegaBytes :: Maybe Integer

Describes the total amount of data that has been transfered in MB's.

currentRateInMegaBytesPerSecond :: Maybe Double

Describes the data transfer rate in MB's per second.

elapsedTimeInSeconds :: Maybe Integer

Describes the number of seconds that have elapsed during the data transfer.

totalDataInMegaBytes :: Maybe Integer

Describes the total amount of data to be transfered in megabytes.

newDataTransferProgress :: DataTransferProgress

Create a value of DataTransferProgress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DataTransferProgress', dataTransferProgress_status - Describes the status of the cluster. While the transfer is in progress the status is transferringdata.

$sel:estimatedTimeToCompletionInSeconds:DataTransferProgress', dataTransferProgress_estimatedTimeToCompletionInSeconds - Describes the estimated number of seconds remaining to complete the transfer.

$sel:dataTransferredInMegaBytes:DataTransferProgress', dataTransferProgress_dataTransferredInMegaBytes - Describes the total amount of data that has been transfered in MB's.

$sel:currentRateInMegaBytesPerSecond:DataTransferProgress', dataTransferProgress_currentRateInMegaBytesPerSecond - Describes the data transfer rate in MB's per second.

$sel:elapsedTimeInSeconds:DataTransferProgress', dataTransferProgress_elapsedTimeInSeconds - Describes the number of seconds that have elapsed during the data transfer.

$sel:totalDataInMegaBytes:DataTransferProgress', dataTransferProgress_totalDataInMegaBytes - Describes the total amount of data to be transfered in megabytes.

dataTransferProgress_status :: Lens' DataTransferProgress (Maybe Text)

Describes the status of the cluster. While the transfer is in progress the status is transferringdata.

dataTransferProgress_estimatedTimeToCompletionInSeconds :: Lens' DataTransferProgress (Maybe Integer)

Describes the estimated number of seconds remaining to complete the transfer.

dataTransferProgress_dataTransferredInMegaBytes :: Lens' DataTransferProgress (Maybe Integer)

Describes the total amount of data that has been transfered in MB's.

dataTransferProgress_currentRateInMegaBytesPerSecond :: Lens' DataTransferProgress (Maybe Double)

Describes the data transfer rate in MB's per second.

dataTransferProgress_elapsedTimeInSeconds :: Lens' DataTransferProgress (Maybe Integer)

Describes the number of seconds that have elapsed during the data transfer.

dataTransferProgress_totalDataInMegaBytes :: Lens' DataTransferProgress (Maybe Integer)

Describes the total amount of data to be transfered in megabytes.