Network.AWS.Redshift.Types.EventSubscription

Description

Documentation

data EventSubscription

Describes event subscriptions.

See: newEventSubscription smart constructor.

Constructors

EventSubscription'

Fields

custSubscriptionId :: Maybe Text

The name of the Amazon Redshift event notification subscription.

status :: Maybe Text

The status of the Amazon Redshift event notification subscription.

Constraints:

Can be one of the following: active | no-permission | topic-not-exist

The status "no-permission" indicates that Amazon Redshift no longer has permission to post to the Amazon SNS topic. The status "topic-not-exist" indicates that the topic was deleted after the subscription was created.

sourceIdsList :: Maybe [Text]

A list of the sources that publish events to the Amazon Redshift event notification subscription.

severity :: Maybe Text

The event severity specified in the Amazon Redshift event notification subscription.

Values: ERROR, INFO

eventCategoriesList :: Maybe [Text]

The list of Amazon Redshift event categories specified in the event notification subscription.

Values: Configuration, Management, Monitoring, Security, Pending

enabled :: Maybe Bool

A boolean value indicating whether the subscription is enabled; true indicates that the subscription is enabled.

subscriptionCreationTime :: Maybe ISO8601

The date and time the Amazon Redshift event notification subscription was created.

customerAwsId :: Maybe Text

The Amazon Web Services account associated with the Amazon Redshift event notification subscription.

tags :: Maybe [Tag]

The list of tags for the event subscription.

sourceType :: Maybe Text

The source type of the events returned by the Amazon Redshift event notification, such as cluster, cluster-snapshot, cluster-parameter-group, cluster-security-group, or scheduled-action.

snsTopicArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon SNS topic used by the event notification subscription.

newEventSubscription :: EventSubscription

Create a value of EventSubscription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:custSubscriptionId:EventSubscription', eventSubscription_custSubscriptionId - The name of the Amazon Redshift event notification subscription.

$sel:status:EventSubscription', eventSubscription_status - The status of the Amazon Redshift event notification subscription.

Constraints:

Can be one of the following: active | no-permission | topic-not-exist

The status "no-permission" indicates that Amazon Redshift no longer has permission to post to the Amazon SNS topic. The status "topic-not-exist" indicates that the topic was deleted after the subscription was created.

$sel:sourceIdsList:EventSubscription', eventSubscription_sourceIdsList - A list of the sources that publish events to the Amazon Redshift event notification subscription.

$sel:severity:EventSubscription', eventSubscription_severity - The event severity specified in the Amazon Redshift event notification subscription.

Values: ERROR, INFO

$sel:eventCategoriesList:EventSubscription', eventSubscription_eventCategoriesList - The list of Amazon Redshift event categories specified in the event notification subscription.

Values: Configuration, Management, Monitoring, Security, Pending

$sel:enabled:EventSubscription', eventSubscription_enabled - A boolean value indicating whether the subscription is enabled; true indicates that the subscription is enabled.

$sel:subscriptionCreationTime:EventSubscription', eventSubscription_subscriptionCreationTime - The date and time the Amazon Redshift event notification subscription was created.

$sel:customerAwsId:EventSubscription', eventSubscription_customerAwsId - The Amazon Web Services account associated with the Amazon Redshift event notification subscription.

$sel:tags:EventSubscription', eventSubscription_tags - The list of tags for the event subscription.

$sel:sourceType:EventSubscription', eventSubscription_sourceType - The source type of the events returned by the Amazon Redshift event notification, such as cluster, cluster-snapshot, cluster-parameter-group, cluster-security-group, or scheduled-action.

$sel:snsTopicArn:EventSubscription', eventSubscription_snsTopicArn - The Amazon Resource Name (ARN) of the Amazon SNS topic used by the event notification subscription.

eventSubscription_custSubscriptionId :: Lens' EventSubscription (Maybe Text)

The name of the Amazon Redshift event notification subscription.

eventSubscription_status :: Lens' EventSubscription (Maybe Text)

The status of the Amazon Redshift event notification subscription.

Constraints:

Can be one of the following: active | no-permission | topic-not-exist

The status "no-permission" indicates that Amazon Redshift no longer has permission to post to the Amazon SNS topic. The status "topic-not-exist" indicates that the topic was deleted after the subscription was created.

eventSubscription_sourceIdsList :: Lens' EventSubscription (Maybe [Text])

A list of the sources that publish events to the Amazon Redshift event notification subscription.

eventSubscription_severity :: Lens' EventSubscription (Maybe Text)

The event severity specified in the Amazon Redshift event notification subscription.

Values: ERROR, INFO

eventSubscription_eventCategoriesList :: Lens' EventSubscription (Maybe [Text])

The list of Amazon Redshift event categories specified in the event notification subscription.

Values: Configuration, Management, Monitoring, Security, Pending

eventSubscription_enabled :: Lens' EventSubscription (Maybe Bool)

A boolean value indicating whether the subscription is enabled; true indicates that the subscription is enabled.

eventSubscription_subscriptionCreationTime :: Lens' EventSubscription (Maybe UTCTime)

The date and time the Amazon Redshift event notification subscription was created.

eventSubscription_customerAwsId :: Lens' EventSubscription (Maybe Text)

The Amazon Web Services account associated with the Amazon Redshift event notification subscription.

eventSubscription_tags :: Lens' EventSubscription (Maybe [Tag])

The list of tags for the event subscription.

eventSubscription_sourceType :: Lens' EventSubscription (Maybe Text)

The source type of the events returned by the Amazon Redshift event notification, such as cluster, cluster-snapshot, cluster-parameter-group, cluster-security-group, or scheduled-action.

eventSubscription_snsTopicArn :: Lens' EventSubscription (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon SNS topic used by the event notification subscription.