Network.AWS.Redshift.Types.NodeConfigurationOptionsFilter

Description

Documentation

data NodeConfigurationOptionsFilter

A set of elements to filter the returned node configurations.

See: newNodeConfigurationOptionsFilter smart constructor.

Constructors

NodeConfigurationOptionsFilter'

Fields

values :: Maybe [Text]

List of values. Compare Name using Operator to Values. If filter Name is NumberOfNodes, then values can range from 0 to 200. If filter Name is EstimatedDiskUtilizationPercent, then values can range from 0 to 100. For example, filter NumberOfNodes (name) GT (operator) 3 (values).

operator :: Maybe OperatorType

The filter operator. If filter Name is NodeType only the 'in' operator is supported. Provide one value to evaluate for 'eq', 'lt', 'le', 'gt', and 'ge'. Provide two values to evaluate for 'between'. Provide a list of values for 'in'.

name :: Maybe NodeConfigurationOptionsFilterName

The name of the element to filter.

newNodeConfigurationOptionsFilter :: NodeConfigurationOptionsFilter

Create a value of NodeConfigurationOptionsFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:NodeConfigurationOptionsFilter', nodeConfigurationOptionsFilter_values - List of values. Compare Name using Operator to Values. If filter Name is NumberOfNodes, then values can range from 0 to 200. If filter Name is EstimatedDiskUtilizationPercent, then values can range from 0 to 100. For example, filter NumberOfNodes (name) GT (operator) 3 (values).

$sel:operator:NodeConfigurationOptionsFilter', nodeConfigurationOptionsFilter_operator - The filter operator. If filter Name is NodeType only the 'in' operator is supported. Provide one value to evaluate for 'eq', 'lt', 'le', 'gt', and 'ge'. Provide two values to evaluate for 'between'. Provide a list of values for 'in'.

$sel:name:NodeConfigurationOptionsFilter', nodeConfigurationOptionsFilter_name - The name of the element to filter.

nodeConfigurationOptionsFilter_values :: Lens' NodeConfigurationOptionsFilter (Maybe [Text])

List of values. Compare Name using Operator to Values. If filter Name is NumberOfNodes, then values can range from 0 to 200. If filter Name is EstimatedDiskUtilizationPercent, then values can range from 0 to 100. For example, filter NumberOfNodes (name) GT (operator) 3 (values).

nodeConfigurationOptionsFilter_operator :: Lens' NodeConfigurationOptionsFilter (Maybe OperatorType)

The filter operator. If filter Name is NodeType only the 'in' operator is supported. Provide one value to evaluate for 'eq', 'lt', 'le', 'gt', and 'ge'. Provide two values to evaluate for 'between'. Provide a list of values for 'in'.

nodeConfigurationOptionsFilter_name :: Lens' NodeConfigurationOptionsFilter (Maybe NodeConfigurationOptionsFilterName)

The name of the element to filter.