Network.AWS.Redshift.Types.OrderableClusterOption

Description

Documentation

data OrderableClusterOption

Describes an orderable cluster option.

See: newOrderableClusterOption smart constructor.

Constructors

OrderableClusterOption'

Fields

availabilityZones :: Maybe [AvailabilityZone]

A list of availability zones for the orderable cluster.

clusterType :: Maybe Text

The cluster type, for example multi-node.

clusterVersion :: Maybe Text

The version of the orderable cluster.

nodeType :: Maybe Text

The node type for the orderable cluster.

newOrderableClusterOption :: OrderableClusterOption

Create a value of OrderableClusterOption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityZones:OrderableClusterOption', orderableClusterOption_availabilityZones - A list of availability zones for the orderable cluster.

$sel:clusterType:OrderableClusterOption', orderableClusterOption_clusterType - The cluster type, for example multi-node.

$sel:clusterVersion:OrderableClusterOption', orderableClusterOption_clusterVersion - The version of the orderable cluster.

$sel:nodeType:OrderableClusterOption', orderableClusterOption_nodeType - The node type for the orderable cluster.

orderableClusterOption_availabilityZones :: Lens' OrderableClusterOption (Maybe [AvailabilityZone])

A list of availability zones for the orderable cluster.

orderableClusterOption_clusterType :: Lens' OrderableClusterOption (Maybe Text)

The cluster type, for example multi-node.

orderableClusterOption_clusterVersion :: Lens' OrderableClusterOption (Maybe Text)

The version of the orderable cluster.

orderableClusterOption_nodeType :: Lens' OrderableClusterOption (Maybe Text)

The node type for the orderable cluster.