Network.AWS.Redshift.Types.Parameter

Description

Documentation

data Parameter

Describes a parameter in a cluster parameter group.

See: newParameter smart constructor.

Constructors

Parameter'

Fields

allowedValues :: Maybe Text

The valid range of values for the parameter.

source :: Maybe Text

The source of the parameter value, such as "engine-default" or "user".

parameterValue :: Maybe Text

The value of the parameter. If ParameterName is wlm_json_configuration, then the maximum size of ParameterValue is 8000 characters.

applyType :: Maybe ParameterApplyType

Specifies how to apply the WLM configuration parameter. Some properties can be applied dynamically, while other properties require that any associated clusters be rebooted for the configuration changes to be applied. For more information about parameters and parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Cluster Management Guide.

parameterName :: Maybe Text

The name of the parameter.

description :: Maybe Text

A description of the parameter.

dataType :: Maybe Text

The data type of the parameter.

isModifiable :: Maybe Bool

If true, the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

minimumEngineVersion :: Maybe Text

The earliest engine version to which the parameter can apply.

newParameter :: Parameter

Create a value of Parameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:allowedValues:Parameter', parameter_allowedValues - The valid range of values for the parameter.

$sel:source:Parameter', parameter_source - The source of the parameter value, such as "engine-default" or "user".

$sel:parameterValue:Parameter', parameter_parameterValue - The value of the parameter. If ParameterName is wlm_json_configuration, then the maximum size of ParameterValue is 8000 characters.

$sel:applyType:Parameter', parameter_applyType - Specifies how to apply the WLM configuration parameter. Some properties can be applied dynamically, while other properties require that any associated clusters be rebooted for the configuration changes to be applied. For more information about parameters and parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Cluster Management Guide.

$sel:parameterName:Parameter', parameter_parameterName - The name of the parameter.

$sel:description:Parameter', parameter_description - A description of the parameter.

$sel:dataType:Parameter', parameter_dataType - The data type of the parameter.

$sel:isModifiable:Parameter', parameter_isModifiable - If true, the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

$sel:minimumEngineVersion:Parameter', parameter_minimumEngineVersion - The earliest engine version to which the parameter can apply.

parameter_allowedValues :: Lens' Parameter (Maybe Text)

The valid range of values for the parameter.

parameter_source :: Lens' Parameter (Maybe Text)

The source of the parameter value, such as "engine-default" or "user".

parameter_parameterValue :: Lens' Parameter (Maybe Text)

The value of the parameter. If ParameterName is wlm_json_configuration, then the maximum size of ParameterValue is 8000 characters.

parameter_applyType :: Lens' Parameter (Maybe ParameterApplyType)

Specifies how to apply the WLM configuration parameter. Some properties can be applied dynamically, while other properties require that any associated clusters be rebooted for the configuration changes to be applied. For more information about parameters and parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Cluster Management Guide.

parameter_parameterName :: Lens' Parameter (Maybe Text)

The name of the parameter.

parameter_description :: Lens' Parameter (Maybe Text)

A description of the parameter.

parameter_dataType :: Lens' Parameter (Maybe Text)

The data type of the parameter.

parameter_isModifiable :: Lens' Parameter (Maybe Bool)

If true, the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

parameter_minimumEngineVersion :: Lens' Parameter (Maybe Text)

The earliest engine version to which the parameter can apply.