Network.AWS.Redshift.Types.PartnerIntegrationInfo

Description

Documentation

data PartnerIntegrationInfo

Describes a partner integration.

See: newPartnerIntegrationInfo smart constructor.

Constructors

PartnerIntegrationInfo'

Fields

status :: Maybe PartnerIntegrationStatus

The partner integration status.

createdAt :: Maybe ISO8601

The date (UTC) that the partner integration was created.

partnerName :: Maybe Text

The name of the partner.

statusMessage :: Maybe Text

The status message provided by the partner.

databaseName :: Maybe Text

The name of the database that receives data from a partner.

updatedAt :: Maybe ISO8601

The date (UTC) that the partner integration status was last updated by the partner.

newPartnerIntegrationInfo :: PartnerIntegrationInfo

Create a value of PartnerIntegrationInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:PartnerIntegrationInfo', partnerIntegrationInfo_status - The partner integration status.

$sel:createdAt:PartnerIntegrationInfo', partnerIntegrationInfo_createdAt - The date (UTC) that the partner integration was created.

$sel:partnerName:PartnerIntegrationInfo', partnerIntegrationInfo_partnerName - The name of the partner.

$sel:statusMessage:PartnerIntegrationInfo', partnerIntegrationInfo_statusMessage - The status message provided by the partner.

$sel:databaseName:PartnerIntegrationInfo', partnerIntegrationInfo_databaseName - The name of the database that receives data from a partner.

$sel:updatedAt:PartnerIntegrationInfo', partnerIntegrationInfo_updatedAt - The date (UTC) that the partner integration status was last updated by the partner.

partnerIntegrationInfo_status :: Lens' PartnerIntegrationInfo (Maybe PartnerIntegrationStatus)

The partner integration status.

partnerIntegrationInfo_createdAt :: Lens' PartnerIntegrationInfo (Maybe UTCTime)

The date (UTC) that the partner integration was created.

partnerIntegrationInfo_partnerName :: Lens' PartnerIntegrationInfo (Maybe Text)

The name of the partner.

partnerIntegrationInfo_statusMessage :: Lens' PartnerIntegrationInfo (Maybe Text)

The status message provided by the partner.

partnerIntegrationInfo_databaseName :: Lens' PartnerIntegrationInfo (Maybe Text)

The name of the database that receives data from a partner.

partnerIntegrationInfo_updatedAt :: Lens' PartnerIntegrationInfo (Maybe UTCTime)

The date (UTC) that the partner integration status was last updated by the partner.