Network.AWS.Redshift.Types.PendingModifiedValues

Description

Documentation

data PendingModifiedValues

Describes cluster attributes that are in a pending state. A change to one or more the attributes was requested and is in progress or will be applied.

See: newPendingModifiedValues smart constructor.

Constructors

PendingModifiedValues'

Fields

encryptionType :: Maybe Text

The encryption type for a cluster. Possible values are: KMS and None.

enhancedVpcRouting :: Maybe Bool

An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

masterUserPassword :: Maybe Text

The pending or in-progress change of the admin user password for the cluster.

publiclyAccessible :: Maybe Bool

The pending or in-progress change of the ability to connect to the cluster from the public network.

maintenanceTrackName :: Maybe Text

The name of the maintenance track that the cluster will change to during the next maintenance window.

automatedSnapshotRetentionPeriod :: Maybe Int

The pending or in-progress change of the automated snapshot retention period.

clusterIdentifier :: Maybe Text

The pending or in-progress change of the new identifier for the cluster.

numberOfNodes :: Maybe Int

The pending or in-progress change of the number of nodes in the cluster.

clusterType :: Maybe Text

The pending or in-progress change of the cluster type.

clusterVersion :: Maybe Text

The pending or in-progress change of the service version.

nodeType :: Maybe Text

The pending or in-progress change of the cluster's node type.

newPendingModifiedValues :: PendingModifiedValues

Create a value of PendingModifiedValues with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionType:PendingModifiedValues', pendingModifiedValues_encryptionType - The encryption type for a cluster. Possible values are: KMS and None.

$sel:enhancedVpcRouting:PendingModifiedValues', pendingModifiedValues_enhancedVpcRouting - An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

$sel:masterUserPassword:PendingModifiedValues', pendingModifiedValues_masterUserPassword - The pending or in-progress change of the admin user password for the cluster.

$sel:publiclyAccessible:PendingModifiedValues', pendingModifiedValues_publiclyAccessible - The pending or in-progress change of the ability to connect to the cluster from the public network.

$sel:maintenanceTrackName:PendingModifiedValues', pendingModifiedValues_maintenanceTrackName - The name of the maintenance track that the cluster will change to during the next maintenance window.

$sel:automatedSnapshotRetentionPeriod:PendingModifiedValues', pendingModifiedValues_automatedSnapshotRetentionPeriod - The pending or in-progress change of the automated snapshot retention period.

$sel:clusterIdentifier:PendingModifiedValues', pendingModifiedValues_clusterIdentifier - The pending or in-progress change of the new identifier for the cluster.

$sel:numberOfNodes:PendingModifiedValues', pendingModifiedValues_numberOfNodes - The pending or in-progress change of the number of nodes in the cluster.

$sel:clusterType:PendingModifiedValues', pendingModifiedValues_clusterType - The pending or in-progress change of the cluster type.

$sel:clusterVersion:PendingModifiedValues', pendingModifiedValues_clusterVersion - The pending or in-progress change of the service version.

$sel:nodeType:PendingModifiedValues', pendingModifiedValues_nodeType - The pending or in-progress change of the cluster's node type.

pendingModifiedValues_encryptionType :: Lens' PendingModifiedValues (Maybe Text)

The encryption type for a cluster. Possible values are: KMS and None.

pendingModifiedValues_enhancedVpcRouting :: Lens' PendingModifiedValues (Maybe Bool)

An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

pendingModifiedValues_masterUserPassword :: Lens' PendingModifiedValues (Maybe Text)

The pending or in-progress change of the admin user password for the cluster.

pendingModifiedValues_publiclyAccessible :: Lens' PendingModifiedValues (Maybe Bool)

The pending or in-progress change of the ability to connect to the cluster from the public network.

pendingModifiedValues_maintenanceTrackName :: Lens' PendingModifiedValues (Maybe Text)

The name of the maintenance track that the cluster will change to during the next maintenance window.

pendingModifiedValues_automatedSnapshotRetentionPeriod :: Lens' PendingModifiedValues (Maybe Int)

The pending or in-progress change of the automated snapshot retention period.

pendingModifiedValues_clusterIdentifier :: Lens' PendingModifiedValues (Maybe Text)

The pending or in-progress change of the new identifier for the cluster.

pendingModifiedValues_numberOfNodes :: Lens' PendingModifiedValues (Maybe Int)

The pending or in-progress change of the number of nodes in the cluster.

pendingModifiedValues_clusterType :: Lens' PendingModifiedValues (Maybe Text)

The pending or in-progress change of the cluster type.

pendingModifiedValues_clusterVersion :: Lens' PendingModifiedValues (Maybe Text)

The pending or in-progress change of the service version.

pendingModifiedValues_nodeType :: Lens' PendingModifiedValues (Maybe Text)

The pending or in-progress change of the cluster's node type.