Network.AWS.Redshift.Types.ReservedNode

Description

Documentation

data ReservedNode

Describes a reserved node. You can call the DescribeReservedNodeOfferings API to obtain the available reserved node offerings.

See: newReservedNode smart constructor.

Constructors

ReservedNode'

Fields

reservedNodeOfferingType :: Maybe ReservedNodeOfferingType

reservedNodeId :: Maybe Text

The unique identifier for the reservation.

reservedNodeOfferingId :: Maybe Text

The identifier for the reserved node offering.

duration :: Maybe Int

The duration of the node reservation in seconds.

startTime :: Maybe ISO8601

The time the reservation started. You purchase a reserved node offering for a duration. This is the start time of that duration.

currencyCode :: Maybe Text

The currency code for the reserved cluster.

state :: Maybe Text

The state of the reserved compute node.

Possible Values:

pending-payment-This reserved node has recently been purchased, and the sale has been approved, but payment has not yet been confirmed.

active-This reserved node is owned by the caller and is available for use.

payment-failed-Payment failed for the purchase attempt.

retired-The reserved node is no longer available.

exchanging-The owner is exchanging the reserved node for another reserved node.

fixedPrice :: Maybe Double

The fixed cost Amazon Redshift charges you for this reserved node.

nodeCount :: Maybe Int

The number of reserved compute nodes.

usagePrice :: Maybe Double

The hourly rate Amazon Redshift charges you for this reserved node.

offeringType :: Maybe Text

The anticipated utilization of the reserved node, as defined in the reserved node offering.

nodeType :: Maybe Text

The node type of the reserved node.

recurringCharges :: Maybe [RecurringCharge]

The recurring charges for the reserved node.

newReservedNode :: ReservedNode

Create a value of ReservedNode with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservedNodeOfferingType:ReservedNode', reservedNode_reservedNodeOfferingType -

$sel:reservedNodeId:ReservedNode', reservedNode_reservedNodeId - The unique identifier for the reservation.

$sel:reservedNodeOfferingId:ReservedNode', reservedNode_reservedNodeOfferingId - The identifier for the reserved node offering.

$sel:duration:ReservedNode', reservedNode_duration - The duration of the node reservation in seconds.

$sel:startTime:ReservedNode', reservedNode_startTime - The time the reservation started. You purchase a reserved node offering for a duration. This is the start time of that duration.

$sel:currencyCode:ReservedNode', reservedNode_currencyCode - The currency code for the reserved cluster.

$sel:state:ReservedNode', reservedNode_state - The state of the reserved compute node.

Possible Values:

pending-payment-This reserved node has recently been purchased, and the sale has been approved, but payment has not yet been confirmed.

active-This reserved node is owned by the caller and is available for use.

payment-failed-Payment failed for the purchase attempt.

retired-The reserved node is no longer available.

exchanging-The owner is exchanging the reserved node for another reserved node.

$sel:fixedPrice:ReservedNode', reservedNode_fixedPrice - The fixed cost Amazon Redshift charges you for this reserved node.

$sel:nodeCount:ReservedNode', reservedNode_nodeCount - The number of reserved compute nodes.

$sel:usagePrice:ReservedNode', reservedNode_usagePrice - The hourly rate Amazon Redshift charges you for this reserved node.

$sel:offeringType:ReservedNode', reservedNode_offeringType - The anticipated utilization of the reserved node, as defined in the reserved node offering.

$sel:nodeType:ReservedNode', reservedNode_nodeType - The node type of the reserved node.

$sel:recurringCharges:ReservedNode', reservedNode_recurringCharges - The recurring charges for the reserved node.

reservedNode_reservedNodeOfferingType :: Lens' ReservedNode (Maybe ReservedNodeOfferingType)

reservedNode_reservedNodeId :: Lens' ReservedNode (Maybe Text)

The unique identifier for the reservation.

reservedNode_reservedNodeOfferingId :: Lens' ReservedNode (Maybe Text)

The identifier for the reserved node offering.

reservedNode_duration :: Lens' ReservedNode (Maybe Int)

The duration of the node reservation in seconds.

reservedNode_startTime :: Lens' ReservedNode (Maybe UTCTime)

The time the reservation started. You purchase a reserved node offering for a duration. This is the start time of that duration.

reservedNode_currencyCode :: Lens' ReservedNode (Maybe Text)

The currency code for the reserved cluster.

reservedNode_state :: Lens' ReservedNode (Maybe Text)

The state of the reserved compute node.

Possible Values:

pending-payment-This reserved node has recently been purchased, and the sale has been approved, but payment has not yet been confirmed.

active-This reserved node is owned by the caller and is available for use.

payment-failed-Payment failed for the purchase attempt.

retired-The reserved node is no longer available.

exchanging-The owner is exchanging the reserved node for another reserved node.

reservedNode_fixedPrice :: Lens' ReservedNode (Maybe Double)

The fixed cost Amazon Redshift charges you for this reserved node.

reservedNode_nodeCount :: Lens' ReservedNode (Maybe Int)

The number of reserved compute nodes.

reservedNode_usagePrice :: Lens' ReservedNode (Maybe Double)

The hourly rate Amazon Redshift charges you for this reserved node.

reservedNode_offeringType :: Lens' ReservedNode (Maybe Text)

The anticipated utilization of the reserved node, as defined in the reserved node offering.

reservedNode_nodeType :: Lens' ReservedNode (Maybe Text)

The node type of the reserved node.

reservedNode_recurringCharges :: Lens' ReservedNode (Maybe [RecurringCharge])

The recurring charges for the reserved node.