Network.AWS.Redshift.Types.ResizeProgressMessage

Description

Documentation

data ResizeProgressMessage

Describes the result of a cluster resize operation.

See: newResizeProgressMessage smart constructor.

Constructors

ResizeProgressMessage'

Fields

importTablesNotStarted :: Maybe [Text]

The names of tables that have not been yet imported.

Valid Values: List of table names

status :: Maybe Text

The status of the resize operation.

Valid Values: NONE | IN_PROGRESS | FAILED | SUCCEEDED | CANCELLING

estimatedTimeToCompletionInSeconds :: Maybe Integer

The estimated time remaining, in seconds, until the resize operation is complete. This value is calculated based on the average resize rate and the estimated amount of data remaining to be processed. Once the resize operation is complete, this value will be 0.

avgResizeRateInMegaBytesPerSecond :: Maybe Double

The average rate of the resize operation over the last few minutes, measured in megabytes per second. After the resize operation completes, this value shows the average rate of the entire resize operation.

targetNumberOfNodes :: Maybe Int

The number of nodes that the cluster will have after the resize operation is complete.

targetEncryptionType :: Maybe Text

The type of encryption for the cluster after the resize is complete.

Possible values are KMS and None.

targetNodeType :: Maybe Text

The node type that the cluster will have after the resize operation is complete.

importTablesInProgress :: Maybe [Text]

The names of tables that are being currently imported.

Valid Values: List of table names.

resizeType :: Maybe Text

An enum with possible values of ClassicResize and ElasticResize. These values describe the type of resize operation being performed.

importTablesCompleted :: Maybe [Text]

The names of tables that have been completely imported .

Valid Values: List of table names.

progressInMegaBytes :: Maybe Integer

While the resize operation is in progress, this value shows the current amount of data, in megabytes, that has been processed so far. When the resize operation is complete, this value shows the total amount of data, in megabytes, on the cluster, which may be more or less than TotalResizeDataInMegaBytes (the estimated total amount of data before resize).

dataTransferProgressPercent :: Maybe Double

The percent of data transferred from source cluster to target cluster.

totalResizeDataInMegaBytes :: Maybe Integer

The estimated total amount of data, in megabytes, on the cluster before the resize operation began.

targetClusterType :: Maybe Text

The cluster type after the resize operation is complete.

Valid Values: multi-node | single-node

message :: Maybe Text

An optional string to provide additional details about the resize action.

elapsedTimeInSeconds :: Maybe Integer

The amount of seconds that have elapsed since the resize operation began. After the resize operation completes, this value shows the total actual time, in seconds, for the resize operation.

newResizeProgressMessage :: ResizeProgressMessage

Create a value of ResizeProgressMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:importTablesNotStarted:ResizeProgressMessage', resizeProgressMessage_importTablesNotStarted - The names of tables that have not been yet imported.

Valid Values: List of table names

$sel:status:ResizeProgressMessage', resizeProgressMessage_status - The status of the resize operation.

Valid Values: NONE | IN_PROGRESS | FAILED | SUCCEEDED | CANCELLING

$sel:estimatedTimeToCompletionInSeconds:ResizeProgressMessage', resizeProgressMessage_estimatedTimeToCompletionInSeconds - The estimated time remaining, in seconds, until the resize operation is complete. This value is calculated based on the average resize rate and the estimated amount of data remaining to be processed. Once the resize operation is complete, this value will be 0.

$sel:avgResizeRateInMegaBytesPerSecond:ResizeProgressMessage', resizeProgressMessage_avgResizeRateInMegaBytesPerSecond - The average rate of the resize operation over the last few minutes, measured in megabytes per second. After the resize operation completes, this value shows the average rate of the entire resize operation.

$sel:targetNumberOfNodes:ResizeProgressMessage', resizeProgressMessage_targetNumberOfNodes - The number of nodes that the cluster will have after the resize operation is complete.

$sel:targetEncryptionType:ResizeProgressMessage', resizeProgressMessage_targetEncryptionType - The type of encryption for the cluster after the resize is complete.

Possible values are KMS and None.

$sel:targetNodeType:ResizeProgressMessage', resizeProgressMessage_targetNodeType - The node type that the cluster will have after the resize operation is complete.

$sel:importTablesInProgress:ResizeProgressMessage', resizeProgressMessage_importTablesInProgress - The names of tables that are being currently imported.

Valid Values: List of table names.

$sel:resizeType:ResizeProgressMessage', resizeProgressMessage_resizeType - An enum with possible values of ClassicResize and ElasticResize. These values describe the type of resize operation being performed.

$sel:importTablesCompleted:ResizeProgressMessage', resizeProgressMessage_importTablesCompleted - The names of tables that have been completely imported .

Valid Values: List of table names.

$sel:progressInMegaBytes:ResizeProgressMessage', resizeProgressMessage_progressInMegaBytes - While the resize operation is in progress, this value shows the current amount of data, in megabytes, that has been processed so far. When the resize operation is complete, this value shows the total amount of data, in megabytes, on the cluster, which may be more or less than TotalResizeDataInMegaBytes (the estimated total amount of data before resize).

$sel:dataTransferProgressPercent:ResizeProgressMessage', resizeProgressMessage_dataTransferProgressPercent - The percent of data transferred from source cluster to target cluster.

$sel:totalResizeDataInMegaBytes:ResizeProgressMessage', resizeProgressMessage_totalResizeDataInMegaBytes - The estimated total amount of data, in megabytes, on the cluster before the resize operation began.

$sel:targetClusterType:ResizeProgressMessage', resizeProgressMessage_targetClusterType - The cluster type after the resize operation is complete.

Valid Values: multi-node | single-node

$sel:message:ResizeProgressMessage', resizeProgressMessage_message - An optional string to provide additional details about the resize action.

$sel:elapsedTimeInSeconds:ResizeProgressMessage', resizeProgressMessage_elapsedTimeInSeconds - The amount of seconds that have elapsed since the resize operation began. After the resize operation completes, this value shows the total actual time, in seconds, for the resize operation.

resizeProgressMessage_importTablesNotStarted :: Lens' ResizeProgressMessage (Maybe [Text])

The names of tables that have not been yet imported.

Valid Values: List of table names

resizeProgressMessage_status :: Lens' ResizeProgressMessage (Maybe Text)

The status of the resize operation.

Valid Values: NONE | IN_PROGRESS | FAILED | SUCCEEDED | CANCELLING

resizeProgressMessage_estimatedTimeToCompletionInSeconds :: Lens' ResizeProgressMessage (Maybe Integer)

The estimated time remaining, in seconds, until the resize operation is complete. This value is calculated based on the average resize rate and the estimated amount of data remaining to be processed. Once the resize operation is complete, this value will be 0.

resizeProgressMessage_avgResizeRateInMegaBytesPerSecond :: Lens' ResizeProgressMessage (Maybe Double)

The average rate of the resize operation over the last few minutes, measured in megabytes per second. After the resize operation completes, this value shows the average rate of the entire resize operation.

resizeProgressMessage_targetNumberOfNodes :: Lens' ResizeProgressMessage (Maybe Int)

The number of nodes that the cluster will have after the resize operation is complete.

resizeProgressMessage_targetEncryptionType :: Lens' ResizeProgressMessage (Maybe Text)

The type of encryption for the cluster after the resize is complete.

Possible values are KMS and None.

resizeProgressMessage_targetNodeType :: Lens' ResizeProgressMessage (Maybe Text)

The node type that the cluster will have after the resize operation is complete.

resizeProgressMessage_importTablesInProgress :: Lens' ResizeProgressMessage (Maybe [Text])

The names of tables that are being currently imported.

Valid Values: List of table names.

resizeProgressMessage_resizeType :: Lens' ResizeProgressMessage (Maybe Text)

An enum with possible values of ClassicResize and ElasticResize. These values describe the type of resize operation being performed.

resizeProgressMessage_importTablesCompleted :: Lens' ResizeProgressMessage (Maybe [Text])

The names of tables that have been completely imported .

Valid Values: List of table names.

resizeProgressMessage_progressInMegaBytes :: Lens' ResizeProgressMessage (Maybe Integer)

While the resize operation is in progress, this value shows the current amount of data, in megabytes, that has been processed so far. When the resize operation is complete, this value shows the total amount of data, in megabytes, on the cluster, which may be more or less than TotalResizeDataInMegaBytes (the estimated total amount of data before resize).

resizeProgressMessage_dataTransferProgressPercent :: Lens' ResizeProgressMessage (Maybe Double)

The percent of data transferred from source cluster to target cluster.

resizeProgressMessage_totalResizeDataInMegaBytes :: Lens' ResizeProgressMessage (Maybe Integer)

The estimated total amount of data, in megabytes, on the cluster before the resize operation began.

resizeProgressMessage_targetClusterType :: Lens' ResizeProgressMessage (Maybe Text)

The cluster type after the resize operation is complete.

Valid Values: multi-node | single-node

resizeProgressMessage_message :: Lens' ResizeProgressMessage (Maybe Text)

An optional string to provide additional details about the resize action.

resizeProgressMessage_elapsedTimeInSeconds :: Lens' ResizeProgressMessage (Maybe Integer)

The amount of seconds that have elapsed since the resize operation began. After the resize operation completes, this value shows the total actual time, in seconds, for the resize operation.