Network.AWS.Redshift.Types.ScheduledAction

Description

Documentation

data ScheduledAction

Describes a scheduled action. You can use a scheduled action to trigger some Amazon Redshift API operations on a schedule. For information about which API operations can be scheduled, see ScheduledActionType.

See: newScheduledAction smart constructor.

Constructors

ScheduledAction'

Fields

state :: Maybe ScheduledActionState

The state of the scheduled action. For example, DISABLED.

targetAction :: Maybe ScheduledActionType

A JSON format string of the Amazon Redshift API operation with input parameters.

"{\"ResizeCluster\":{\"NodeType\":\"ds2.8xlarge\",\"ClusterIdentifier\":\"my-test-cluster\",\"NumberOfNodes\":3}}".

startTime :: Maybe ISO8601

The start time in UTC when the schedule is active. Before this time, the scheduled action does not trigger.

schedule :: Maybe Text

The schedule for a one-time (at format) or recurring (cron format) scheduled action. Schedule invocations must be separated by at least one hour.

Format of at expressions is "at(yyyy-mm-ddThh:mm:ss)". For example, "at(2016-03-04T17:27:00)".

Format of cron expressions is "cron(Minutes Hours Day-of-month Month Day-of-week Year)". For example, "cron(0 10 ? * MON *)". For more information, see Cron Expressions in the Amazon CloudWatch Events User Guide.

scheduledActionName :: Maybe Text

The name of the scheduled action.

scheduledActionDescription :: Maybe Text

The description of the scheduled action.

nextInvocations :: Maybe [ISO8601]

List of times when the scheduled action will run.

endTime :: Maybe ISO8601

The end time in UTC when the schedule is no longer active. After this time, the scheduled action does not trigger.

iamRole :: Maybe Text

The IAM role to assume to run the scheduled action. This IAM role must have permission to run the Amazon Redshift API operation in the scheduled action. This IAM role must allow the Amazon Redshift scheduler (Principal scheduler.redshift.amazonaws.com) to assume permissions on your behalf. For more information about the IAM role to use with the Amazon Redshift scheduler, see Using Identity-Based Policies for Amazon Redshift in the Amazon Redshift Cluster Management Guide.

newScheduledAction :: ScheduledAction

Create a value of ScheduledAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:ScheduledAction', scheduledAction_state - The state of the scheduled action. For example, DISABLED.

$sel:targetAction:ScheduledAction', scheduledAction_targetAction - A JSON format string of the Amazon Redshift API operation with input parameters.

"{\"ResizeCluster\":{\"NodeType\":\"ds2.8xlarge\",\"ClusterIdentifier\":\"my-test-cluster\",\"NumberOfNodes\":3}}".

$sel:startTime:ScheduledAction', scheduledAction_startTime - The start time in UTC when the schedule is active. Before this time, the scheduled action does not trigger.

$sel:schedule:ScheduledAction', scheduledAction_schedule - The schedule for a one-time (at format) or recurring (cron format) scheduled action. Schedule invocations must be separated by at least one hour.

Format of at expressions is "at(yyyy-mm-ddThh:mm:ss)". For example, "at(2016-03-04T17:27:00)".

Format of cron expressions is "cron(Minutes Hours Day-of-month Month Day-of-week Year)". For example, "cron(0 10 ? * MON *)". For more information, see Cron Expressions in the Amazon CloudWatch Events User Guide.

$sel:scheduledActionName:ScheduledAction', scheduledAction_scheduledActionName - The name of the scheduled action.

$sel:scheduledActionDescription:ScheduledAction', scheduledAction_scheduledActionDescription - The description of the scheduled action.

$sel:nextInvocations:ScheduledAction', scheduledAction_nextInvocations - List of times when the scheduled action will run.

$sel:endTime:ScheduledAction', scheduledAction_endTime - The end time in UTC when the schedule is no longer active. After this time, the scheduled action does not trigger.

$sel:iamRole:ScheduledAction', scheduledAction_iamRole - The IAM role to assume to run the scheduled action. This IAM role must have permission to run the Amazon Redshift API operation in the scheduled action. This IAM role must allow the Amazon Redshift scheduler (Principal scheduler.redshift.amazonaws.com) to assume permissions on your behalf. For more information about the IAM role to use with the Amazon Redshift scheduler, see Using Identity-Based Policies for Amazon Redshift in the Amazon Redshift Cluster Management Guide.

scheduledAction_state :: Lens' ScheduledAction (Maybe ScheduledActionState)

The state of the scheduled action. For example, DISABLED.

scheduledAction_targetAction :: Lens' ScheduledAction (Maybe ScheduledActionType)

A JSON format string of the Amazon Redshift API operation with input parameters.

"{\"ResizeCluster\":{\"NodeType\":\"ds2.8xlarge\",\"ClusterIdentifier\":\"my-test-cluster\",\"NumberOfNodes\":3}}".

scheduledAction_startTime :: Lens' ScheduledAction (Maybe UTCTime)

The start time in UTC when the schedule is active. Before this time, the scheduled action does not trigger.

scheduledAction_schedule :: Lens' ScheduledAction (Maybe Text)

The schedule for a one-time (at format) or recurring (cron format) scheduled action. Schedule invocations must be separated by at least one hour.

Format of at expressions is "at(yyyy-mm-ddThh:mm:ss)". For example, "at(2016-03-04T17:27:00)".

Format of cron expressions is "cron(Minutes Hours Day-of-month Month Day-of-week Year)". For example, "cron(0 10 ? * MON *)". For more information, see Cron Expressions in the Amazon CloudWatch Events User Guide.

scheduledAction_scheduledActionName :: Lens' ScheduledAction (Maybe Text)

The name of the scheduled action.

scheduledAction_scheduledActionDescription :: Lens' ScheduledAction (Maybe Text)

The description of the scheduled action.

scheduledAction_nextInvocations :: Lens' ScheduledAction (Maybe [UTCTime])

List of times when the scheduled action will run.

scheduledAction_endTime :: Lens' ScheduledAction (Maybe UTCTime)

The end time in UTC when the schedule is no longer active. After this time, the scheduled action does not trigger.

scheduledAction_iamRole :: Lens' ScheduledAction (Maybe Text)

The IAM role to assume to run the scheduled action. This IAM role must have permission to run the Amazon Redshift API operation in the scheduled action. This IAM role must allow the Amazon Redshift scheduler (Principal scheduler.redshift.amazonaws.com) to assume permissions on your behalf. For more information about the IAM role to use with the Amazon Redshift scheduler, see Using Identity-Based Policies for Amazon Redshift in the Amazon Redshift Cluster Management Guide.