Network.AWS.Redshift.Types.SnapshotCopyGrant

Description

Documentation

data SnapshotCopyGrant

The snapshot copy grant that grants Amazon Redshift permission to encrypt copied snapshots with the specified customer master key (CMK) from Amazon Web Services KMS in the destination region.

For more information about managing snapshot copy grants, go to Amazon Redshift Database Encryption in the Amazon Redshift Cluster Management Guide.

See: newSnapshotCopyGrant smart constructor.

Constructors

SnapshotCopyGrant'

Fields

snapshotCopyGrantName :: Maybe Text

The name of the snapshot copy grant.

kmsKeyId :: Maybe Text

The unique identifier of the customer master key (CMK) in Amazon Web Services KMS to which Amazon Redshift is granted permission.

tags :: Maybe [Tag]

A list of tag instances.

newSnapshotCopyGrant :: SnapshotCopyGrant

Create a value of SnapshotCopyGrant with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshotCopyGrantName:SnapshotCopyGrant', snapshotCopyGrant_snapshotCopyGrantName - The name of the snapshot copy grant.

$sel:kmsKeyId:SnapshotCopyGrant', snapshotCopyGrant_kmsKeyId - The unique identifier of the customer master key (CMK) in Amazon Web Services KMS to which Amazon Redshift is granted permission.

$sel:tags:SnapshotCopyGrant', snapshotCopyGrant_tags - A list of tag instances.

snapshotCopyGrant_snapshotCopyGrantName :: Lens' SnapshotCopyGrant (Maybe Text)

The name of the snapshot copy grant.

snapshotCopyGrant_kmsKeyId :: Lens' SnapshotCopyGrant (Maybe Text)

The unique identifier of the customer master key (CMK) in Amazon Web Services KMS to which Amazon Redshift is granted permission.

snapshotCopyGrant_tags :: Lens' SnapshotCopyGrant (Maybe [Tag])

A list of tag instances.