Network.AWS.Redshift.Types.SnapshotErrorMessage

Description

Documentation

data SnapshotErrorMessage

Describes the errors returned by a snapshot.

See: newSnapshotErrorMessage smart constructor.

Constructors

SnapshotErrorMessage'

Fields

snapshotIdentifier :: Maybe Text

A unique identifier for the snapshot returning the error.

failureCode :: Maybe Text

The failure code for the error.

snapshotClusterIdentifier :: Maybe Text

A unique identifier for the cluster.

failureReason :: Maybe Text

The text message describing the error.

newSnapshotErrorMessage :: SnapshotErrorMessage

Create a value of SnapshotErrorMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshotIdentifier:SnapshotErrorMessage', snapshotErrorMessage_snapshotIdentifier - A unique identifier for the snapshot returning the error.

$sel:failureCode:SnapshotErrorMessage', snapshotErrorMessage_failureCode - The failure code for the error.

$sel:snapshotClusterIdentifier:SnapshotErrorMessage', snapshotErrorMessage_snapshotClusterIdentifier - A unique identifier for the cluster.

$sel:failureReason:SnapshotErrorMessage', snapshotErrorMessage_failureReason - The text message describing the error.

snapshotErrorMessage_snapshotIdentifier :: Lens' SnapshotErrorMessage (Maybe Text)

A unique identifier for the snapshot returning the error.

snapshotErrorMessage_failureCode :: Lens' SnapshotErrorMessage (Maybe Text)

The failure code for the error.

snapshotErrorMessage_snapshotClusterIdentifier :: Lens' SnapshotErrorMessage (Maybe Text)

A unique identifier for the cluster.

snapshotErrorMessage_failureReason :: Lens' SnapshotErrorMessage (Maybe Text)

The text message describing the error.