Network.AWS.Redshift.Types.SnapshotSchedule

Description

Documentation

data SnapshotSchedule

Describes a snapshot schedule. You can set a regular interval for creating snapshots of a cluster. You can also schedule snapshots for specific dates.

See: newSnapshotSchedule smart constructor.

Constructors

SnapshotSchedule'

Fields

associatedClusters :: Maybe [ClusterAssociatedToSchedule]

A list of clusters associated with the schedule. A maximum of 100 clusters is returned.

nextInvocations :: Maybe [ISO8601]

scheduleDefinitions :: Maybe [Text]

A list of ScheduleDefinitions.

scheduleDescription :: Maybe Text

The description of the schedule.

scheduleIdentifier :: Maybe Text

A unique identifier for the schedule.

associatedClusterCount :: Maybe Int

The number of clusters associated with the schedule.

tags :: Maybe [Tag]

An optional set of tags describing the schedule.

newSnapshotSchedule :: SnapshotSchedule

Create a value of SnapshotSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associatedClusters:SnapshotSchedule', snapshotSchedule_associatedClusters - A list of clusters associated with the schedule. A maximum of 100 clusters is returned.

$sel:nextInvocations:SnapshotSchedule', snapshotSchedule_nextInvocations -

$sel:scheduleDefinitions:SnapshotSchedule', snapshotSchedule_scheduleDefinitions - A list of ScheduleDefinitions.

$sel:scheduleDescription:SnapshotSchedule', snapshotSchedule_scheduleDescription - The description of the schedule.

$sel:scheduleIdentifier:SnapshotSchedule', snapshotSchedule_scheduleIdentifier - A unique identifier for the schedule.

$sel:associatedClusterCount:SnapshotSchedule', snapshotSchedule_associatedClusterCount - The number of clusters associated with the schedule.

$sel:tags:SnapshotSchedule', snapshotSchedule_tags - An optional set of tags describing the schedule.

snapshotSchedule_associatedClusters :: Lens' SnapshotSchedule (Maybe [ClusterAssociatedToSchedule])

A list of clusters associated with the schedule. A maximum of 100 clusters is returned.

snapshotSchedule_nextInvocations :: Lens' SnapshotSchedule (Maybe [UTCTime])

snapshotSchedule_scheduleDefinitions :: Lens' SnapshotSchedule (Maybe [Text])

A list of ScheduleDefinitions.

snapshotSchedule_scheduleDescription :: Lens' SnapshotSchedule (Maybe Text)

The description of the schedule.

snapshotSchedule_scheduleIdentifier :: Lens' SnapshotSchedule (Maybe Text)

A unique identifier for the schedule.

snapshotSchedule_associatedClusterCount :: Lens' SnapshotSchedule (Maybe Int)

The number of clusters associated with the schedule.

snapshotSchedule_tags :: Lens' SnapshotSchedule (Maybe [Tag])

An optional set of tags describing the schedule.