Network.AWS.Redshift.Types.TableRestoreStatus

Description

Documentation

data TableRestoreStatus

Describes the status of a RestoreTableFromClusterSnapshot operation.

See: newTableRestoreStatus smart constructor.

Constructors

TableRestoreStatus'

Fields

targetSchemaName :: Maybe Text

The name of the schema to restore the table to.

snapshotIdentifier :: Maybe Text

The identifier of the snapshot that the table is being restored from.

sourceDatabaseName :: Maybe Text

The name of the source database that contains the table being restored.

status :: Maybe TableRestoreStatusType

A value that describes the current state of the table restore request.

Valid Values: SUCCEEDED, FAILED, CANCELED, PENDING, IN_PROGRESS

sourceTableName :: Maybe Text

The name of the source table being restored.

requestTime :: Maybe ISO8601

The time that the table restore request was made, in Universal Coordinated Time (UTC).

sourceSchemaName :: Maybe Text

The name of the source schema that contains the table being restored.

targetDatabaseName :: Maybe Text

The name of the database to restore the table to.

message :: Maybe Text

A description of the status of the table restore request. Status values include SUCCEEDED, FAILED, CANCELED, PENDING, IN_PROGRESS.

clusterIdentifier :: Maybe Text

The identifier of the Amazon Redshift cluster that the table is being restored to.

newTableName' :: Maybe Text

The name of the table to create as a result of the table restore request.

progressInMegaBytes :: Maybe Integer

The amount of data restored to the new table so far, in megabytes (MB).

tableRestoreRequestId :: Maybe Text

The unique identifier for the table restore request.

totalDataInMegaBytes :: Maybe Integer

The total amount of data to restore to the new table, in megabytes (MB).

newTableRestoreStatus :: TableRestoreStatus

Create a value of TableRestoreStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetSchemaName:TableRestoreStatus', tableRestoreStatus_targetSchemaName - The name of the schema to restore the table to.

$sel:snapshotIdentifier:TableRestoreStatus', tableRestoreStatus_snapshotIdentifier - The identifier of the snapshot that the table is being restored from.

$sel:sourceDatabaseName:TableRestoreStatus', tableRestoreStatus_sourceDatabaseName - The name of the source database that contains the table being restored.

$sel:status:TableRestoreStatus', tableRestoreStatus_status - A value that describes the current state of the table restore request.

Valid Values: SUCCEEDED, FAILED, CANCELED, PENDING, IN_PROGRESS

$sel:sourceTableName:TableRestoreStatus', tableRestoreStatus_sourceTableName - The name of the source table being restored.

$sel:requestTime:TableRestoreStatus', tableRestoreStatus_requestTime - The time that the table restore request was made, in Universal Coordinated Time (UTC).

$sel:sourceSchemaName:TableRestoreStatus', tableRestoreStatus_sourceSchemaName - The name of the source schema that contains the table being restored.

$sel:targetDatabaseName:TableRestoreStatus', tableRestoreStatus_targetDatabaseName - The name of the database to restore the table to.

$sel:message:TableRestoreStatus', tableRestoreStatus_message - A description of the status of the table restore request. Status values include SUCCEEDED, FAILED, CANCELED, PENDING, IN_PROGRESS.

$sel:clusterIdentifier:TableRestoreStatus', tableRestoreStatus_clusterIdentifier - The identifier of the Amazon Redshift cluster that the table is being restored to.

$sel:newTableName':TableRestoreStatus', tableRestoreStatus_newTableName - The name of the table to create as a result of the table restore request.

$sel:progressInMegaBytes:TableRestoreStatus', tableRestoreStatus_progressInMegaBytes - The amount of data restored to the new table so far, in megabytes (MB).

$sel:tableRestoreRequestId:TableRestoreStatus', tableRestoreStatus_tableRestoreRequestId - The unique identifier for the table restore request.

$sel:totalDataInMegaBytes:TableRestoreStatus', tableRestoreStatus_totalDataInMegaBytes - The total amount of data to restore to the new table, in megabytes (MB).

tableRestoreStatus_targetSchemaName :: Lens' TableRestoreStatus (Maybe Text)

The name of the schema to restore the table to.

tableRestoreStatus_snapshotIdentifier :: Lens' TableRestoreStatus (Maybe Text)

The identifier of the snapshot that the table is being restored from.

tableRestoreStatus_sourceDatabaseName :: Lens' TableRestoreStatus (Maybe Text)

The name of the source database that contains the table being restored.

tableRestoreStatus_status :: Lens' TableRestoreStatus (Maybe TableRestoreStatusType)

A value that describes the current state of the table restore request.

Valid Values: SUCCEEDED, FAILED, CANCELED, PENDING, IN_PROGRESS

tableRestoreStatus_sourceTableName :: Lens' TableRestoreStatus (Maybe Text)

The name of the source table being restored.

tableRestoreStatus_requestTime :: Lens' TableRestoreStatus (Maybe UTCTime)

The time that the table restore request was made, in Universal Coordinated Time (UTC).

tableRestoreStatus_sourceSchemaName :: Lens' TableRestoreStatus (Maybe Text)

The name of the source schema that contains the table being restored.

tableRestoreStatus_targetDatabaseName :: Lens' TableRestoreStatus (Maybe Text)

The name of the database to restore the table to.

tableRestoreStatus_message :: Lens' TableRestoreStatus (Maybe Text)

A description of the status of the table restore request. Status values include SUCCEEDED, FAILED, CANCELED, PENDING, IN_PROGRESS.

tableRestoreStatus_clusterIdentifier :: Lens' TableRestoreStatus (Maybe Text)

The identifier of the Amazon Redshift cluster that the table is being restored to.

tableRestoreStatus_newTableName :: Lens' TableRestoreStatus (Maybe Text)

The name of the table to create as a result of the table restore request.

tableRestoreStatus_progressInMegaBytes :: Lens' TableRestoreStatus (Maybe Integer)

The amount of data restored to the new table so far, in megabytes (MB).

tableRestoreStatus_tableRestoreRequestId :: Lens' TableRestoreStatus (Maybe Text)

The unique identifier for the table restore request.

tableRestoreStatus_totalDataInMegaBytes :: Lens' TableRestoreStatus (Maybe Integer)

The total amount of data to restore to the new table, in megabytes (MB).