Network.AWS.Redshift.V2012_12_01.AuthorizeSnapshotAccess

Description

Authorizes the specified AWS customer account to restore the specified snapshot. For more information about working with snapshots, go to Amazon Redshift Snapshots in the Amazon Redshift Management Guide.

Documentation

authorizeSnapshotAccess

Arguments

:: Text
_asamSnapshotIdentifier

-> Text
_asamAccountWithRestoreAccess

-> AuthorizeSnapshotAccess

Minimum specification for a AuthorizeSnapshotAccess request.

data AuthorizeSnapshotAccess

Constructors

AuthorizeSnapshotAccess

Fields

_asamSnapshotIdentifier :: Text

The identifier of the snapshot the account is authorized to restore.

_asamAccountWithRestoreAccess :: Text

The identifier of the AWS customer account authorized to restore the specified snapshot.

_asamSnapshotClusterIdentifier :: Maybe Text

The identifier of the cluster the snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

asamSnapshotIdentifier :: Lens' AuthorizeSnapshotAccess Text

asamSnapshotClusterIdentifier :: Lens' AuthorizeSnapshotAccess (Maybe Text)

asamAccountWithRestoreAccess :: Lens' AuthorizeSnapshotAccess Text

data AuthorizeSnapshotAccessResponse

Constructors

AuthorizeSnapshotAccessResponse

Fields

_sssssssssssssrSnapshot :: Maybe Snapshot

Describes a snapshot.

sssssssssssssrSnapshot :: Iso' AuthorizeSnapshotAccessResponse (Maybe Snapshot)