Network.AWS.Redshift.V2012_12_01.CopyClusterSnapshot

Description

Copies the specified automated cluster snapshot to a new manual cluster snapshot. The source must be an automated snapshot and it must be in the available state. When you delete a cluster, Amazon Redshift deletes any automated snapshots of the cluster. Also, when the retention period of the snapshot expires, Amazon Redshift automatically deletes it. If you want to keep an automated snapshot for a longer period, you can make a manual copy of the snapshot. Manual snapshots are retained until you delete them. For more information about working with snapshots, go to Amazon Redshift Snapshots in the Amazon Redshift Management Guide. https:/redshift.us-east-1.amazonaws.com ?Action=CopyClusterSnapshot &SourceSnapshotIdentifier=cm:examplecluster-2013-01-22-19-27-58 &TargetSnapshotIdentifier=my-snapshot-456 &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20130123us-east-1redshiftaws4_request &x-amz-date=20130123T014618Z &x-amz-signedheaders=content-type;host;x-amz-date 5439 my-snapshot-456 available manual 1.0 2013-01-22T19:27:58.931Z 2 dev 2013-01-22T19:23:59.368Z us-east-1c dw1.xlarge examplecluster adminuser aebb56f5-64fe-11e2-88c5-53eb05787dfb.

Documentation

copyClusterSnapshot

Arguments

:: Text
_ccsnTargetSnapshotIdentifier

-> Text
_ccsnSourceSnapshotIdentifier

-> CopyClusterSnapshot

Minimum specification for a CopyClusterSnapshot request.

data CopyClusterSnapshot

Constructors

CopyClusterSnapshot

Fields

_ccsnTargetSnapshotIdentifier :: Text

The identifier given to the new manual snapshot. Constraints: Cannot be null, empty, or blank. Must contain from 1 to 255 alphanumeric characters or hyphens. First character must be a letter. Cannot end with a hyphen or contain two consecutive hyphens. Must be unique for the AWS account that is making the request.

_ccsnSourceSnapshotIdentifier :: Text

The identifier for the source snapshot. Constraints: Must be the identifier for a valid automated snapshot whose state is available.

_ccsnSourceSnapshotClusterIdentifier :: Maybe Text

The identifier of the cluster the source snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name. Constraints: Must be the identifier for a valid cluster.

ccsnTargetSnapshotIdentifier :: Lens' CopyClusterSnapshot Text

ccsnSourceSnapshotIdentifier :: Lens' CopyClusterSnapshot Text

ccsnSourceSnapshotClusterIdentifier :: Lens' CopyClusterSnapshot (Maybe Text)

data CopyClusterSnapshotResponse

Constructors

CopyClusterSnapshotResponse

Fields

_ssssssssssssssssssssssssssssssssssssrSnapshot :: Maybe Snapshot

Describes a snapshot.

ssssssssssssssssssssssssssssssssssssrSnapshot :: Iso' CopyClusterSnapshotResponse (Maybe Snapshot)