Network.AWS.Redshift.V2012_12_01.CreateClusterSnapshot

Description

Creates a manual snapshot of the specified cluster. The cluster must be in the available state. For more information about working with snapshots, go to Amazon Redshift Snapshots in the Amazon Redshift Management Guide. https:/redshift.us-east-1.amazonaws.com ?Action=CreateClusterSnapshot &ClusterIdentifier=examplecluster &SnapshotIdentifier=snapshot-1234 &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20130123us-east-1redshiftaws4_request &x-amz-date=20130123T010824Z &x-amz-signedheaders=content-type;host;x-amz-date 5439 my-snapshot-123 creating manual 1.0 2013-01-23T01:08:29.142Z 2 dev 2013-01-22T19:23:59.368Z us-east-1c dw1.xlarge examplecluster adminuser 65baef14-64f9-11e2-bea9-49e0ce183f07.

Documentation

data CreateClusterSnapshot

Constructors

CreateClusterSnapshot

Fields

_ccsmSnapshotIdentifier :: Text

A unique identifier for the snapshot that you are requesting. This identifier must be unique for all snapshots within the AWS account. Constraints: Cannot be null, empty, or blank Must contain from 1 to 255 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens Example: my-snapshot-id.

_ccsmClusterIdentifier :: Text

The cluster identifier for which you want a snapshot.

ccsmSnapshotIdentifier :: Lens' CreateClusterSnapshot Text

ccsmClusterIdentifier :: Lens' CreateClusterSnapshot Text

data CreateClusterSnapshotResponse

Constructors

CreateClusterSnapshotResponse

Fields

_ssssssssssssssssssssssssssrSnapshot :: Maybe Snapshot

Describes a snapshot.

ssssssssssssssssssssssssssrSnapshot :: Iso' CreateClusterSnapshotResponse (Maybe Snapshot)