Network.AWS.Redshift.V2012_12_01.CreateHsmConfiguration

Description

Creates an HSM configuration that contains the information required by an Amazon Redshift cluster to store and use database encryption keys in a Hardware Security Module (HSM). After creating the HSM configuration, you can specify it as a parameter when creating a cluster. The cluster will then store its encryption keys in the HSM. In addition to creating an HSM configuration, you must also create an HSM client certificate. For more information, go to Hardware Security Modules in the Amazon Redshift Management Guide.

Documentation

data CreateHsmConfiguration

Constructors

CreateHsmConfiguration

Fields

_chcmHsmServerPublicCertificate :: Text

The HSMs public certificate file. When using Cloud HSM, the file name is server.pem.

_chcmHsmConfigurationIdentifier :: Text

The identifier to be assigned to the new Amazon Redshift HSM configuration.

_chcmHsmPartitionName :: Text

The name of the partition in the HSM where the Amazon Redshift clusters will store their database encryption keys.

_chcmHsmPartitionPassword :: Text

The password required to access the HSM partition.

_chcmDescription :: Text

A text description of the HSM configuration to be created.

_chcmHsmIpAddress :: Text

The IP address that the Amazon Redshift cluster must use to access the HSM.

chcmHsmServerPublicCertificate :: Lens' CreateHsmConfiguration Text

chcmHsmPartitionPassword :: Lens' CreateHsmConfiguration Text

chcmHsmPartitionName :: Lens' CreateHsmConfiguration Text

chcmHsmIpAddress :: Lens' CreateHsmConfiguration Text

chcmHsmConfigurationIdentifier :: Lens' CreateHsmConfiguration Text

chcmDescription :: Lens' CreateHsmConfiguration Text

data CreateHsmConfigurationResponse

Constructors

CreateHsmConfigurationResponse

Fields

_hcwHsmConfiguration :: Maybe HsmConfiguration

Returns information about an HSM configuration, which is an object that describes to Amazon Redshift clusters the information they require to connect to an HSM where they can store database encryption keys.

hcwHsmConfiguration :: Iso' CreateHsmConfigurationResponse (Maybe HsmConfiguration)