Network.AWS.Redshift.V2012_12_01.DeleteClusterSnapshot

Description

Deletes the specified manual snapshot. The snapshot must be in the available state, with no other users authorized to access the snapshot. Unlike automated snapshots, manual snapshots are retained even after you delete your cluster. Amazon Redshift does not delete your manual snapshots. You must delete manual snapshot explicitly to avoid getting charged. If other accounts are authorized to access the snapshot, you must revoke all of the authorizations before you can delete the snapshot. https:/redshift.us-east-1.amazonaws.com ?Action=DeleteClusterSnapshot &SnapshotIdentifier=snapshot-1234 &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20121208us-east-1redshiftaws4_request &x-amz-date=20121208T005225Z &x-amz-signedheaders=content-type;host;x-amz-date 2012-12-07T23:31:02.372Z 5439 snapshot-1234 deleted 2012-12-06T23:09:01.475Z manual 1.0 us-east-1a examplecluster masteruser dw1.xlarge mydb 3 88a31de4-40d1-11e2-8a25-eb010998df4e.

Documentation

deleteClusterSnapshot

Arguments

:: Text
_dcsmSnapshotIdentifier

-> DeleteClusterSnapshot

Minimum specification for a DeleteClusterSnapshot request.

data DeleteClusterSnapshot

Constructors

DeleteClusterSnapshot

Fields

_dcsmSnapshotIdentifier :: Text

The unique identifier of the manual snapshot to be deleted. Constraints: Must be the name of an existing snapshot that is in the available state.

_dcsmSnapshotClusterIdentifier :: Maybe Text

The unique identifier of the cluster the snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name. Constraints: Must be the name of valid cluster.

dcsmSnapshotIdentifier :: Lens' DeleteClusterSnapshot Text

dcsmSnapshotClusterIdentifier :: Lens' DeleteClusterSnapshot (Maybe Text)

data DeleteClusterSnapshotResponse

Constructors

DeleteClusterSnapshotResponse

Fields

_sssrSnapshot :: Maybe Snapshot

Describes a snapshot.

sssrSnapshot :: Iso' DeleteClusterSnapshotResponse (Maybe Snapshot)