Network.AWS.Redshift.V2012_12_01.DescribeClusterParameters

Description

Returns a detailed list of parameters contained within the specified Amazon Redshift parameter group. For each parameter the response includes information such as parameter name, description, data type, value, whether the parameter value is modifiable, and so on. You can specify source filter to retrieve parameters of only specific type. For example, to retrieve parameters that were modified by a user action such as from ModifyClusterParameterGroup, you can specify source equal to user. For more information about managing parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Management Guide. https:/redshift.us-east-1.amazonaws.com ?Action=DescribeClusterParameters &ParameterGroupName=parametergroup1 &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20121208us-east-1redshiftaws4_request &x-amz-date=20121208T010408Z &x-amz-signedheaders=content-type;host;x-amz-date ISO, MDY string engine-default true Sets the display format for date and time values. datestyle 0 integer engine-default true Sets the number of digits displayed for floating-point values -15-2 extra_float_digits default string engine-default true This parameter applies a user-defined label to a group of queries that are run during the same session.. query_group false boolean engine-default true require ssl for all databaseconnections true,false require_ssl $user, public string engine-default true Sets the schema search order for names that are not schema-qualified. search_path 0 integer engine-default true Aborts any statement that takes over the specified number of milliseconds. statement_timeout [{&quot;query_concurrency&quot;:5}] string engine-default true wlm json configuration wlm_json_configuration 2ba35df4-40d3-11e2-82cf-0b45b05c0221.

Documentation

describeClusterParameters

Arguments

:: Text
_dcpmParameterGroupName

-> DescribeClusterParameters

Minimum specification for a DescribeClusterParameters request.

data DescribeClusterParameters

Constructors

DescribeClusterParameters

Fields

_dcpmParameterGroupName :: Text

The name of a cluster parameter group for which to return details.

_dcpmMaxRecords :: Maybe Integer

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

_dcpmMarker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterParameters request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

_dcpmSource :: Maybe Text

The parameter types to return. Specify user to show parameters that are different form the default. Similarly, specify engine-default to show parameters that are the same as the default parameter group. Default: All parameter types returned. Valid Values: user | engine-default.

dcpmSource :: Lens' DescribeClusterParameters (Maybe Text)

dcpmParameterGroupName :: Lens' DescribeClusterParameters Text

dcpmMaxRecords :: Lens' DescribeClusterParameters (Maybe Integer)

dcpmMarker :: Lens' DescribeClusterParameters (Maybe Text)

data DescribeClusterParametersResponse

Constructors

DescribeClusterParametersResponse

Fields

_cpgdParameters :: [Parameter]

A list of Parameter instances. Each instance lists the parameters of one cluster parameter group.

_cpgdMarker :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

cpgdParameters :: Lens' DescribeClusterParametersResponse [Parameter]

cpgdMarker :: Lens' DescribeClusterParametersResponse (Maybe Text)