Network.AWS.Redshift.V2012_12_01.DescribeHsmClientCertificates

Description

Returns information about the specified HSM client certificate. If no certificate ID is specified, returns information about all the HSM certificates owned by your AWS customer account.

Documentation

describeHsmClientCertificates :: DescribeHsmClientCertificates

Minimum specification for a DescribeHsmClientCertificates request.

data DescribeHsmClientCertificates

Constructors

DescribeHsmClientCertificates

Fields

_dhccmMaxRecords :: Maybe Integer

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

_dhccmHsmClientCertificateIdentifier :: Maybe Text

The identifier of a specific HSM client certificate for which you want information. If no identifier is specified, information is returned for all HSM client certificates owned by your AWS customer account.

_dhccmMarker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeHsmClientCertificates request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

dhccmMaxRecords :: Lens' DescribeHsmClientCertificates (Maybe Integer)

dhccmMarker :: Lens' DescribeHsmClientCertificates (Maybe Text)

dhccmHsmClientCertificateIdentifier :: Lens' DescribeHsmClientCertificates (Maybe Text)

data DescribeHsmClientCertificatesResponse

Constructors

DescribeHsmClientCertificatesResponse

Fields

_hccmHsmClientCertificates :: [HsmClientCertificate]

A list of the identifiers for one or more HSM client certificates used by Amazon Redshift clusters to store and retrieve database encryption keys in an HSM.

_hccmMarker :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

hccmMarker :: Lens' DescribeHsmClientCertificatesResponse (Maybe Text)

hccmHsmClientCertificates :: Lens' DescribeHsmClientCertificatesResponse [HsmClientCertificate]