Network.AWS.Redshift.V2012_12_01.DescribeHsmConfigurations

Description

Returns information about the specified Amazon Redshift HSM configuration. If no configuration ID is specified, returns information about all the HSM configurations owned by your AWS customer account.

Documentation

describeHsmConfigurations :: DescribeHsmConfigurations

Minimum specification for a DescribeHsmConfigurations request.

data DescribeHsmConfigurations

Constructors

DescribeHsmConfigurations

Fields

_dhcnMaxRecords :: Maybe Integer

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

_dhcnHsmConfigurationIdentifier :: Maybe Text

The identifier of a specific Amazon Redshift HSM configuration to be described. If no identifier is specified, information is returned for all HSM configurations owned by your AWS customer account.

_dhcnMarker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeHsmConfigurations request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

dhcnMaxRecords :: Lens' DescribeHsmConfigurations (Maybe Integer)

dhcnMarker :: Lens' DescribeHsmConfigurations (Maybe Text)

dhcnHsmConfigurationIdentifier :: Lens' DescribeHsmConfigurations (Maybe Text)

data DescribeHsmConfigurationsResponse

Constructors

DescribeHsmConfigurationsResponse

Fields

_hcmHsmConfigurations :: [HsmConfiguration]

A list of Amazon Redshift HSM configurations.

_hcmMarker :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

hcmMarker :: Lens' DescribeHsmConfigurationsResponse (Maybe Text)

hcmHsmConfigurations :: Lens' DescribeHsmConfigurationsResponse [HsmConfiguration]