Network.AWS.Redshift.V2012_12_01.DisableSnapshotCopy

Description

Disables the automatic copying of snapshots from one region to another region for a specified cluster.

Documentation

data DisableSnapshotCopy

Constructors

DisableSnapshotCopy

Fields

_dscmClusterIdentifier :: Text

The unique identifier of the source cluster that you want to disable copying of snapshots to a destination region. Constraints: Must be the valid name of an existing cluster that has cross-region snapshot copy enabled.

dscmClusterIdentifier :: Iso' DisableSnapshotCopy Text

data DisableSnapshotCopyResponse

Constructors

DisableSnapshotCopyResponse

Fields

_ccvCluster :: Maybe Cluster

Describes a cluster.

ccvCluster :: Iso' DisableSnapshotCopyResponse (Maybe Cluster)