Network.AWS.Redshift.V2012_12_01.ModifyClusterSubnetGroup

Description

Modifies a cluster subnet group to include the specified list of VPC subnets. The operation replaces the existing list of subnets with the new list of subnets.

Documentation

modifyClusterSubnetGroup

Arguments

:: Text
_mcsgmClusterSubnetGroupName

-> [Text]
_mcsgmSubnetIds

-> ModifyClusterSubnetGroup

Minimum specification for a ModifyClusterSubnetGroup request.

data ModifyClusterSubnetGroup

Constructors

ModifyClusterSubnetGroup

Fields

_mcsgmClusterSubnetGroupName :: Text

The name of the subnet group to be modified.

_mcsgmSubnetIds :: [Text]

An array of VPC subnet IDs. A maximum of 20 subnets can be modified in a single request.

_mcsgmDescription :: Maybe Text

A text description of the subnet group to be modified.

mcsgmSubnetIds :: Lens' ModifyClusterSubnetGroup [Text]

mcsgmDescription :: Lens' ModifyClusterSubnetGroup (Maybe Text)

mcsgmClusterSubnetGroupName :: Lens' ModifyClusterSubnetGroup Text

data ModifyClusterSubnetGroupResponse

Constructors

ModifyClusterSubnetGroupResponse

Fields

_csgcrClusterSubnetGroup :: Maybe ClusterSubnetGroup

Describes a subnet group.

csgcrClusterSubnetGroup :: Iso' ModifyClusterSubnetGroupResponse (Maybe ClusterSubnetGroup)