Network.AWS.Redshift.V2012_12_01.ModifyEventSubscription

Description

Modifies an existing Amazon Redshift event notification subscription.

Documentation

modifyEventSubscription

Arguments

:: Text
_mesmSubscriptionName

-> ModifyEventSubscription

Minimum specification for a ModifyEventSubscription request.

data ModifyEventSubscription

Constructors

ModifyEventSubscription

Fields

_mesmSubscriptionName :: Text

The name of the modified Amazon Redshift event notification subscription.

_mesmEnabled :: Maybe Bool

A Boolean value indicating if the subscription is enabled. true indicates the subscription is enabled.

_mesmEventCategories :: [Text]

Specifies the Amazon Redshift event categories to be published by the event notification subscription. Values: Configuration, Management, Monitoring, Security.

_mesmSourceIds :: [Text]

A list of one or more identifiers of Amazon Redshift source objects. All of the objects must be of the same type as was specified in the source type parameter. The event subscription will return only events generated by the specified objects. If not specified, then events are returned for all objects within the source type specified. Example: my-cluster-1, my-cluster-2 Example: my-snapshot-20131010.

_mesmSnsTopicArn :: Maybe Text

The Amazon Resource Name (ARN) of the SNS topic to be used by the event notification subscription.

_mesmSourceType :: Maybe Text

The type of source that will be generating the events. For example, if you want to be notified of events generated by a cluster, you would set this parameter to cluster. If this value is not specified, events are returned for all Amazon Redshift objects in your AWS account. You must specify a source type in order to specify source IDs. Valid values: cluster, cluster-parameter-group, cluster-security-group, and cluster-snapshot.

_mesmSeverity :: Maybe Text

Specifies the Amazon Redshift event severity to be published by the event notification subscription. Values: ERROR, INFO.

mesmSubscriptionName :: Lens' ModifyEventSubscription Text

mesmSourceType :: Lens' ModifyEventSubscription (Maybe Text)

mesmSourceIds :: Lens' ModifyEventSubscription [Text]

mesmSnsTopicArn :: Lens' ModifyEventSubscription (Maybe Text)

mesmSeverity :: Lens' ModifyEventSubscription (Maybe Text)

mesmEventCategories :: Lens' ModifyEventSubscription [Text]

mesmEnabled :: Lens' ModifyEventSubscription (Maybe Bool)

data ModifyEventSubscriptionResponse

Constructors

ModifyEventSubscriptionResponse

Fields

_eswEventSubscription :: Maybe EventSubscription

eswEventSubscription :: Iso' ModifyEventSubscriptionResponse (Maybe EventSubscription)