Network.AWS.Redshift.V2012_12_01.PurchaseReservedNodeOffering

Description

Allows you to purchase reserved nodes. Amazon Redshift offers a predefined set of reserved node offerings. You can purchase one of the offerings. You can call the DescribeReservedNodeOfferings API to obtain the available reserved node offerings. You can call this API by providing a specific reserved node offering and the number of nodes you want to reserve. For more information about managing parameter groups, go to Purchasing Reserved Nodes in the Amazon Redshift Management Guide. https:/redshift.us-east-1.amazonaws.com ?Action=PurchaseReservedNodeOffering &ReservedNodeOfferingId=3a98bf7d-979a-49cc-b568-18f24315baf0 &NodeCount=2 &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20130117us-east-1redshiftaws4_request &x-amz-date=20130117T232351Z &x-amz-signedheaders=content-type;host;x-amz-date 2013-01-18T21:42:44.402Z Heavy Utilization 94608000 Hourly 0.21 12452.0 0.0 payment-pending dw1.8xlarge 2 1ba8e2e3-dacf-48d9-841f-cc675182a8a6 fcb117cc-61b7-11e2-b6e9-87e586e4ca38.

Documentation

purchaseReservedNodeOffering

Arguments

:: Text
_prnomReservedNodeOfferingId

-> PurchaseReservedNodeOffering

Minimum specification for a PurchaseReservedNodeOffering request.

data PurchaseReservedNodeOffering

Constructors

PurchaseReservedNodeOffering

Fields

_prnomReservedNodeOfferingId :: Text

The unique identifier of the reserved node offering you want to purchase.

_prnomNodeCount :: Maybe Integer

The number of reserved nodes you want to purchase. Default: 1.

prnomReservedNodeOfferingId :: Lens' PurchaseReservedNodeOffering Text

prnomNodeCount :: Lens' PurchaseReservedNodeOffering (Maybe Integer)

data PurchaseReservedNodeOfferingResponse

Constructors

PurchaseReservedNodeOfferingResponse

Fields

_rnwReservedNode :: Maybe ReservedNode

Describes a reserved node.

rnwReservedNode :: Iso' PurchaseReservedNodeOfferingResponse (Maybe ReservedNode)