Network.AWS.Redshift.V2012_12_01.ResetClusterParameterGroup

Description

Sets one or more parameters of the specified parameter group to their default values and sets the source values of the parameters to "engine-default". To reset the entire parameter group specify the ResetAllParameters parameter. For parameter changes to take effect you must reboot any associated clusters. https:/redshift.us-east-1.amazonaws.com ?Action=ResetClusterParameterGroup &ParameterGroupName=parametergroup1 &Parameters.member.1.ParameterName=extra_float_digits &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20121208us-east-1redshiftaws4_request &x-amz-date=20121208T020847Z &x-amz-signedheaders=content-type;host;x-amz-date Your parameter group has been updated but changes won't get applied until you reboot the associated Clusters. parametergroup1 625d23c1-40dc-11e2-8a25-eb010998df4e.

Documentation

resetClusterParameterGroup

Arguments

:: Text
_rcpgmParameterGroupName

-> ResetClusterParameterGroup

Minimum specification for a ResetClusterParameterGroup request.

data ResetClusterParameterGroup

Constructors

ResetClusterParameterGroup

Fields

_rcpgmParameterGroupName :: Text

The name of the cluster parameter group to be reset.

_rcpgmResetAllParameters :: Maybe Bool

If true, all parameters in the specified parameter group will be reset to their default values. Default: true.

_rcpgmParameters :: [Parameter]

An array of names of parameters to be reset. If ResetAllParameters option is not used, then at least one parameter name must be supplied. Constraints: A maximum of 20 parameters can be reset in a single request.

rcpgmResetAllParameters :: Lens' ResetClusterParameterGroup (Maybe Bool)

rcpgmParameters :: Lens' ResetClusterParameterGroup [Parameter]

rcpgmParameterGroupName :: Lens' ResetClusterParameterGroup Text

data ResetClusterParameterGroupResponse

Constructors

ResetClusterParameterGroupResponse

Fields

_cpgnmParameterGroupStatus :: Maybe Text

The status of the parameter group. For example, if you made a change to a parameter group name-value pair, then the change could be pending a reboot of an associated cluster.

_cpgnmParameterGroupName :: Maybe Text

The name of the cluster parameter group.

cpgnmParameterGroupStatus :: Lens' ResetClusterParameterGroupResponse (Maybe Text)

cpgnmParameterGroupName :: Lens' ResetClusterParameterGroupResponse (Maybe Text)