Network.AWS.Redshift.V2012_12_01.RevokeClusterSecurityGroupIngress

Description

Revokes an ingress rule in an Amazon Redshift security group for a previously authorized IP range or Amazon EC2 security group. To add an ingress rule, see AuthorizeClusterSecurityGroupIngress. For information about managing security groups, go to Amazon Redshift Cluster Security Groups in the Amazon Redshift Management Guide. https:/redshift.us-east-1.amazonaws.com ?Action=RevokeClusterSecurityGroupIngress &ClusterSecurityGroupName=securitygroup1 &CIDRIP=192.168.40.3/32 &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20130123us-east-1redshiftaws4_request &x-amz-date=20130123T021606Z &x-amz-signedheaders=content-type;host;x-amz-date my security group securitygroup1 d8eff363-6502-11e2-a8da-655adc216806.

Documentation

revokeClusterSecurityGroupIngress

Arguments

:: Text
_rcsgimClusterSecurityGroupName

-> RevokeClusterSecurityGroupIngress

Minimum specification for a RevokeClusterSecurityGroupIngress request.

data RevokeClusterSecurityGroupIngress

Constructors

RevokeClusterSecurityGroupIngress

Fields

_rcsgimClusterSecurityGroupName :: Text

The name of the security Group from which to revoke the ingress rule.

_rcsgimEC2SecurityGroupOwnerId :: Maybe Text

The AWS account number of the owner of the security group specified in the EC2SecurityGroupName parameter. The AWS access key ID is not an acceptable value. If EC2SecurityGroupOwnerId is specified, EC2SecurityGroupName must also be provided. and CIDRIP cannot be provided. Example: 111122223333.

_rcsgimEC2SecurityGroupName :: Maybe Text

The name of the EC2 Security Group whose access is to be revoked. If EC2SecurityGroupName is specified, EC2SecurityGroupOwnerId must also be provided and CIDRIP cannot be provided.

_rcsgimCIDRIP :: Maybe Text

The IP range for which to revoke access. This range must be a valid Classless Inter-Domain Routing (CIDR) block of IP addresses. If CIDRIP is specified, EC2SecurityGroupName and EC2SecurityGroupOwnerId cannot be provided.

rcsgimEC2SecurityGroupOwnerId :: Lens' RevokeClusterSecurityGroupIngress (Maybe Text)

rcsgimEC2SecurityGroupName :: Lens' RevokeClusterSecurityGroupIngress (Maybe Text)

rcsgimClusterSecurityGroupName :: Lens' RevokeClusterSecurityGroupIngress Text

rcsgimCIDRIP :: Lens' RevokeClusterSecurityGroupIngress (Maybe Text)

data RevokeClusterSecurityGroupIngressResponse

Constructors

RevokeClusterSecurityGroupIngressResponse

Fields

_csgxClusterSecurityGroup :: Maybe ClusterSecurityGroup

Describes a security group.

csgxClusterSecurityGroup :: Iso' RevokeClusterSecurityGroupIngressResponse (Maybe ClusterSecurityGroup)