Network.AWS.RedshiftData.BatchExecuteStatement

Description

Runs one or more SQL statements, which can be data manipulation language (DML) or data definition language (DDL). Depending on the authorization method, use one of the following combinations of request parameters:

Secrets Manager - specify the Amazon Resource Name (ARN) of the secret, the database name, and the cluster identifier that matches the cluster in the secret.

Temporary credentials - specify the cluster identifier, the database name, and the database user name. Permission to call the redshift:GetClusterCredentials operation is required to use this method.

Creating a Request

data BatchExecuteStatement

See: newBatchExecuteStatement smart constructor.

Constructors

BatchExecuteStatement'

Fields

dbUser :: Maybe Text

The database user name. This parameter is required when authenticating using temporary credentials.

statementName :: Maybe Text

The name of the SQL statements. You can name the SQL statements when you create them to identify the query.

secretArn :: Maybe Text

The name or ARN of the secret that enables access to the database. This parameter is required when authenticating using Secrets Manager.

withEvent :: Maybe Bool

A value that indicates whether to send an event to the Amazon EventBridge event bus after the SQL statements run.

clusterIdentifier :: Text

The cluster identifier. This parameter is required when authenticating using either Secrets Manager or temporary credentials.

database :: Text

The name of the database. This parameter is required when authenticating using either Secrets Manager or temporary credentials.

sqls :: NonEmpty Text

One or more SQL statements to run.

newBatchExecuteStatement

Arguments

:: Text
$sel:clusterIdentifier:BatchExecuteStatement'

-> Text
$sel:database:BatchExecuteStatement'

-> NonEmpty Text
$sel:sqls:BatchExecuteStatement'

-> BatchExecuteStatement

Create a value of BatchExecuteStatement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbUser:BatchExecuteStatement', batchExecuteStatement_dbUser - The database user name. This parameter is required when authenticating using temporary credentials.

$sel:statementName:BatchExecuteStatement', batchExecuteStatement_statementName - The name of the SQL statements. You can name the SQL statements when you create them to identify the query.

$sel:secretArn:BatchExecuteStatement', batchExecuteStatement_secretArn - The name or ARN of the secret that enables access to the database. This parameter is required when authenticating using Secrets Manager.

$sel:withEvent:BatchExecuteStatement', batchExecuteStatement_withEvent - A value that indicates whether to send an event to the Amazon EventBridge event bus after the SQL statements run.

$sel:clusterIdentifier:BatchExecuteStatement', batchExecuteStatement_clusterIdentifier - The cluster identifier. This parameter is required when authenticating using either Secrets Manager or temporary credentials.

$sel:database:BatchExecuteStatement', batchExecuteStatement_database - The name of the database. This parameter is required when authenticating using either Secrets Manager or temporary credentials.

$sel:sqls:BatchExecuteStatement', batchExecuteStatement_sqls - One or more SQL statements to run.

Request Lenses

batchExecuteStatement_dbUser :: Lens' BatchExecuteStatement (Maybe Text)

The database user name. This parameter is required when authenticating using temporary credentials.

batchExecuteStatement_statementName :: Lens' BatchExecuteStatement (Maybe Text)

The name of the SQL statements. You can name the SQL statements when you create them to identify the query.

batchExecuteStatement_secretArn :: Lens' BatchExecuteStatement (Maybe Text)

The name or ARN of the secret that enables access to the database. This parameter is required when authenticating using Secrets Manager.

batchExecuteStatement_withEvent :: Lens' BatchExecuteStatement (Maybe Bool)

A value that indicates whether to send an event to the Amazon EventBridge event bus after the SQL statements run.

batchExecuteStatement_clusterIdentifier :: Lens' BatchExecuteStatement Text

The cluster identifier. This parameter is required when authenticating using either Secrets Manager or temporary credentials.

batchExecuteStatement_database :: Lens' BatchExecuteStatement Text

The name of the database. This parameter is required when authenticating using either Secrets Manager or temporary credentials.

batchExecuteStatement_sqls :: Lens' BatchExecuteStatement (NonEmpty Text)

One or more SQL statements to run.

Destructuring the Response

data BatchExecuteStatementResponse

See: newBatchExecuteStatementResponse smart constructor.

Constructors

BatchExecuteStatementResponse'

Fields

dbUser :: Maybe Text

The database user name.

database :: Maybe Text

The name of the database.

createdAt :: Maybe POSIX

The date and time (UTC) the statement was created.

clusterIdentifier :: Maybe Text

The cluster identifier.

id :: Maybe Text

The identifier of the SQL statement whose results are to be fetched. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API. This identifier is returned by BatchExecuteStatment.

secretArn :: Maybe Text

The name or ARN of the secret that enables access to the database.

httpStatus :: Int

The response's http status code.

newBatchExecuteStatementResponse

Arguments

:: Int
$sel:httpStatus:BatchExecuteStatementResponse'

-> BatchExecuteStatementResponse

Create a value of BatchExecuteStatementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbUser:BatchExecuteStatement', batchExecuteStatementResponse_dbUser - The database user name.

$sel:database:BatchExecuteStatement', batchExecuteStatementResponse_database - The name of the database.

$sel:createdAt:BatchExecuteStatementResponse', batchExecuteStatementResponse_createdAt - The date and time (UTC) the statement was created.

$sel:clusterIdentifier:BatchExecuteStatement', batchExecuteStatementResponse_clusterIdentifier - The cluster identifier.

$sel:id:BatchExecuteStatementResponse', batchExecuteStatementResponse_id - The identifier of the SQL statement whose results are to be fetched. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API. This identifier is returned by BatchExecuteStatment.

$sel:secretArn:BatchExecuteStatement', batchExecuteStatementResponse_secretArn - The name or ARN of the secret that enables access to the database.

$sel:httpStatus:BatchExecuteStatementResponse', batchExecuteStatementResponse_httpStatus - The response's http status code.

Response Lenses

batchExecuteStatementResponse_dbUser :: Lens' BatchExecuteStatementResponse (Maybe Text)

The database user name.

batchExecuteStatementResponse_database :: Lens' BatchExecuteStatementResponse (Maybe Text)

The name of the database.

batchExecuteStatementResponse_createdAt :: Lens' BatchExecuteStatementResponse (Maybe UTCTime)

The date and time (UTC) the statement was created.

batchExecuteStatementResponse_clusterIdentifier :: Lens' BatchExecuteStatementResponse (Maybe Text)

The cluster identifier.

batchExecuteStatementResponse_id :: Lens' BatchExecuteStatementResponse (Maybe Text)

The identifier of the SQL statement whose results are to be fetched. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API. This identifier is returned by BatchExecuteStatment.

batchExecuteStatementResponse_secretArn :: Lens' BatchExecuteStatementResponse (Maybe Text)

The name or ARN of the secret that enables access to the database.

batchExecuteStatementResponse_httpStatus :: Lens' BatchExecuteStatementResponse Int

The response's http status code.