Network.AWS.RedshiftData.Types.StatementData

Description

Documentation

data StatementData

The SQL statement to run.

See: newStatementData smart constructor.

Constructors

StatementData'

Fields

status :: Maybe StatusString

The status of the SQL statement. An example is the that the SQL statement finished.

createdAt :: Maybe POSIX

The date and time (UTC) the statement was created.

queryParameters :: Maybe (NonEmpty SqlParameter)

The parameters used in a SQL statement.

queryStrings :: Maybe [Text]

One or more SQL statements. Each query string in the array corresponds to one of the queries in a batch query request.

queryString :: Maybe Text

The SQL statement.

statementName :: Maybe Text

The name of the SQL statement.

updatedAt :: Maybe POSIX

The date and time (UTC) that the statement metadata was last updated.

secretArn :: Maybe Text

The name or Amazon Resource Name (ARN) of the secret that enables access to the database.

isBatchStatement :: Maybe Bool

A value that indicates whether the statement is a batch query request.

id :: Text

The SQL statement identifier. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API.

newStatementData

Arguments

:: Text
$sel:id:StatementData'

-> StatementData

Create a value of StatementData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:StatementData', statementData_status - The status of the SQL statement. An example is the that the SQL statement finished.

$sel:createdAt:StatementData', statementData_createdAt - The date and time (UTC) the statement was created.

$sel:queryParameters:StatementData', statementData_queryParameters - The parameters used in a SQL statement.

$sel:queryStrings:StatementData', statementData_queryStrings - One or more SQL statements. Each query string in the array corresponds to one of the queries in a batch query request.

$sel:queryString:StatementData', statementData_queryString - The SQL statement.

$sel:statementName:StatementData', statementData_statementName - The name of the SQL statement.

$sel:updatedAt:StatementData', statementData_updatedAt - The date and time (UTC) that the statement metadata was last updated.

$sel:secretArn:StatementData', statementData_secretArn - The name or Amazon Resource Name (ARN) of the secret that enables access to the database.

$sel:isBatchStatement:StatementData', statementData_isBatchStatement - A value that indicates whether the statement is a batch query request.

$sel:id:StatementData', statementData_id - The SQL statement identifier. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API.

statementData_status :: Lens' StatementData (Maybe StatusString)

The status of the SQL statement. An example is the that the SQL statement finished.

statementData_createdAt :: Lens' StatementData (Maybe UTCTime)

The date and time (UTC) the statement was created.

statementData_queryParameters :: Lens' StatementData (Maybe (NonEmpty SqlParameter))

The parameters used in a SQL statement.

statementData_queryStrings :: Lens' StatementData (Maybe [Text])

One or more SQL statements. Each query string in the array corresponds to one of the queries in a batch query request.

statementData_queryString :: Lens' StatementData (Maybe Text)

The SQL statement.

statementData_statementName :: Lens' StatementData (Maybe Text)

The name of the SQL statement.

statementData_updatedAt :: Lens' StatementData (Maybe UTCTime)

The date and time (UTC) that the statement metadata was last updated.

statementData_secretArn :: Lens' StatementData (Maybe Text)

The name or Amazon Resource Name (ARN) of the secret that enables access to the database.

statementData_isBatchStatement :: Lens' StatementData (Maybe Bool)

A value that indicates whether the statement is a batch query request.

statementData_id :: Lens' StatementData Text

The SQL statement identifier. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API.