Network.AWS.RedshiftData.Types.SubStatementData

Description

Documentation

data SubStatementData

Information about an SQL statement.

See: newSubStatementData smart constructor.

Constructors

SubStatementData'

Fields

status :: Maybe StatementStatusString

The status of the SQL statement. An example is the that the SQL statement finished.

redshiftQueryId :: Maybe Integer

The SQL statement identifier. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API.

resultSize :: Maybe Integer

The size in bytes of the returned results. A -1 indicates the value is null.

createdAt :: Maybe POSIX

The date and time (UTC) the statement was created.

error :: Maybe Text

The error message from the cluster if the SQL statement encountered an error while running.

resultRows :: Maybe Integer

Either the number of rows returned from the SQL statement or the number of rows affected. If result size is greater than zero, the result rows can be the number of rows affected by SQL statements such as INSERT, UPDATE, DELETE, COPY, and others. A -1 indicates the value is null.

hasResultSet :: Maybe Bool

A value that indicates whether the statement has a result set. The result set can be empty.

queryString :: Maybe Text

The SQL statement text.

updatedAt :: Maybe POSIX

The date and time (UTC) that the statement metadata was last updated.

duration :: Maybe Integer

The amount of time in nanoseconds that the statement ran.

id :: Text

The identifier of the SQL statement. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API. A suffix indicates the number of the SQL statement. For example, d9b6c0c9-0747-4bf4-b142-e8883122f766:2 has a suffix of :2 that indicates the second SQL statement of a batch query.

newSubStatementData

Arguments

:: Text
$sel:id:SubStatementData'

-> SubStatementData

Create a value of SubStatementData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SubStatementData', subStatementData_status - The status of the SQL statement. An example is the that the SQL statement finished.

$sel:redshiftQueryId:SubStatementData', subStatementData_redshiftQueryId - The SQL statement identifier. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API.

$sel:resultSize:SubStatementData', subStatementData_resultSize - The size in bytes of the returned results. A -1 indicates the value is null.

$sel:createdAt:SubStatementData', subStatementData_createdAt - The date and time (UTC) the statement was created.

$sel:error:SubStatementData', subStatementData_error - The error message from the cluster if the SQL statement encountered an error while running.

$sel:resultRows:SubStatementData', subStatementData_resultRows - Either the number of rows returned from the SQL statement or the number of rows affected. If result size is greater than zero, the result rows can be the number of rows affected by SQL statements such as INSERT, UPDATE, DELETE, COPY, and others. A -1 indicates the value is null.

$sel:hasResultSet:SubStatementData', subStatementData_hasResultSet - A value that indicates whether the statement has a result set. The result set can be empty.

$sel:queryString:SubStatementData', subStatementData_queryString - The SQL statement text.

$sel:updatedAt:SubStatementData', subStatementData_updatedAt - The date and time (UTC) that the statement metadata was last updated.

$sel:duration:SubStatementData', subStatementData_duration - The amount of time in nanoseconds that the statement ran.

$sel:id:SubStatementData', subStatementData_id - The identifier of the SQL statement. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API. A suffix indicates the number of the SQL statement. For example, d9b6c0c9-0747-4bf4-b142-e8883122f766:2 has a suffix of :2 that indicates the second SQL statement of a batch query.

subStatementData_status :: Lens' SubStatementData (Maybe StatementStatusString)

The status of the SQL statement. An example is the that the SQL statement finished.

subStatementData_redshiftQueryId :: Lens' SubStatementData (Maybe Integer)

The SQL statement identifier. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API.

subStatementData_resultSize :: Lens' SubStatementData (Maybe Integer)

The size in bytes of the returned results. A -1 indicates the value is null.

subStatementData_createdAt :: Lens' SubStatementData (Maybe UTCTime)

The date and time (UTC) the statement was created.

subStatementData_error :: Lens' SubStatementData (Maybe Text)

The error message from the cluster if the SQL statement encountered an error while running.

subStatementData_resultRows :: Lens' SubStatementData (Maybe Integer)

Either the number of rows returned from the SQL statement or the number of rows affected. If result size is greater than zero, the result rows can be the number of rows affected by SQL statements such as INSERT, UPDATE, DELETE, COPY, and others. A -1 indicates the value is null.

subStatementData_hasResultSet :: Lens' SubStatementData (Maybe Bool)

A value that indicates whether the statement has a result set. The result set can be empty.

subStatementData_queryString :: Lens' SubStatementData (Maybe Text)

The SQL statement text.

subStatementData_updatedAt :: Lens' SubStatementData (Maybe UTCTime)

The date and time (UTC) that the statement metadata was last updated.

subStatementData_duration :: Lens' SubStatementData (Maybe Integer)

The amount of time in nanoseconds that the statement ran.

subStatementData_id :: Lens' SubStatementData Text

The identifier of the SQL statement. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API. A suffix indicates the number of the SQL statement. For example, d9b6c0c9-0747-4bf4-b142-e8883122f766:2 has a suffix of :2 that indicates the second SQL statement of a batch query.