Network.AWS.Rekognition.CreateCollection

Description

Creates a collection in an AWS Region. You can add faces to the collection using the IndexFaces operation.

For example, you might create collections, one for each of your application users. A user can then index faces using the IndexFaces operation and persist results in a specific collection. Then, a user can search the collection for faces in the user-specific container.

When you create a collection, it is associated with the latest version of the face model version.

Collection names are case-sensitive.

This operation requires permissions to perform the rekognition:CreateCollection action. If you want to tag your collection, you also require permission to perform the rekognition:TagResource operation.

Creating a Request

data CreateCollection

See: newCreateCollection smart constructor.

Constructors

CreateCollection'

Fields

tags :: Maybe (HashMap Text Text)

A set of tags (key-value pairs) that you want to attach to the collection.

collectionId :: Text

ID for the collection that you are creating.

newCreateCollection

Arguments

:: Text
$sel:collectionId:CreateCollection'

-> CreateCollection

Create a value of CreateCollection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateCollection', createCollection_tags - A set of tags (key-value pairs) that you want to attach to the collection.

$sel:collectionId:CreateCollection', createCollection_collectionId - ID for the collection that you are creating.

Request Lenses

createCollection_tags :: Lens' CreateCollection (Maybe (HashMap Text Text))

A set of tags (key-value pairs) that you want to attach to the collection.

createCollection_collectionId :: Lens' CreateCollection Text

ID for the collection that you are creating.

Destructuring the Response

data CreateCollectionResponse

See: newCreateCollectionResponse smart constructor.

Constructors

CreateCollectionResponse'

Fields

faceModelVersion :: Maybe Text

Version number of the face detection model associated with the collection you are creating.

collectionArn :: Maybe Text

Amazon Resource Name (ARN) of the collection. You can use this to manage permissions on your resources.

statusCode :: Maybe Natural

HTTP status code indicating the result of the operation.

httpStatus :: Int

The response's http status code.

newCreateCollectionResponse

Arguments

:: Int
$sel:httpStatus:CreateCollectionResponse'

-> CreateCollectionResponse

Create a value of CreateCollectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:faceModelVersion:CreateCollectionResponse', createCollectionResponse_faceModelVersion - Version number of the face detection model associated with the collection you are creating.

$sel:collectionArn:CreateCollectionResponse', createCollectionResponse_collectionArn - Amazon Resource Name (ARN) of the collection. You can use this to manage permissions on your resources.

$sel:statusCode:CreateCollectionResponse', createCollectionResponse_statusCode - HTTP status code indicating the result of the operation.

$sel:httpStatus:CreateCollectionResponse', createCollectionResponse_httpStatus - The response's http status code.

Response Lenses

createCollectionResponse_faceModelVersion :: Lens' CreateCollectionResponse (Maybe Text)

Version number of the face detection model associated with the collection you are creating.

createCollectionResponse_collectionArn :: Lens' CreateCollectionResponse (Maybe Text)

Amazon Resource Name (ARN) of the collection. You can use this to manage permissions on your resources.

createCollectionResponse_statusCode :: Lens' CreateCollectionResponse (Maybe Natural)

HTTP status code indicating the result of the operation.

createCollectionResponse_httpStatus :: Lens' CreateCollectionResponse Int

The response's http status code.