Network.AWS.Rekognition.CreateProject

Description

Creates a new Amazon Rekognition Custom Labels project. A project is a logical grouping of resources (images, Labels, models) and operations (training, evaluation and detection).

This operation requires permissions to perform the rekognition:CreateProject action.

Creating a Request

data CreateProject

See: newCreateProject smart constructor.

Constructors

CreateProject'

Fields

projectName :: Text

The name of the project to create.

newCreateProject

Arguments

:: Text
$sel:projectName:CreateProject'

-> CreateProject

Create a value of CreateProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectName:CreateProject', createProject_projectName - The name of the project to create.

Request Lenses

createProject_projectName :: Lens' CreateProject Text

The name of the project to create.

Destructuring the Response

data CreateProjectResponse

See: newCreateProjectResponse smart constructor.

Constructors

CreateProjectResponse'

Fields

projectArn :: Maybe Text

The Amazon Resource Name (ARN) of the new project. You can use the ARN to configure IAM access to the project.

httpStatus :: Int

The response's http status code.

newCreateProjectResponse

Arguments

:: Int
$sel:httpStatus:CreateProjectResponse'

-> CreateProjectResponse

Create a value of CreateProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectArn:CreateProjectResponse', createProjectResponse_projectArn - The Amazon Resource Name (ARN) of the new project. You can use the ARN to configure IAM access to the project.

$sel:httpStatus:CreateProjectResponse', createProjectResponse_httpStatus - The response's http status code.

Response Lenses

createProjectResponse_projectArn :: Lens' CreateProjectResponse (Maybe Text)

The Amazon Resource Name (ARN) of the new project. You can use the ARN to configure IAM access to the project.

createProjectResponse_httpStatus :: Lens' CreateProjectResponse Int

The response's http status code.