Network.AWS.Rekognition.CreateProjectVersion

Description

Creates a new version of a model and begins training. Models are managed as part of an Amazon Rekognition Custom Labels project. You can specify one training dataset and one testing dataset. The response from CreateProjectVersion is an Amazon Resource Name (ARN) for the version of the model.

Training takes a while to complete. You can get the current status by calling DescribeProjectVersions.

Once training has successfully completed, call DescribeProjectVersions to get the training results and evaluate the model.

After evaluating the model, you start the model by calling StartProjectVersion.

This operation requires permissions to perform the rekognition:CreateProjectVersion action.

Creating a Request

data CreateProjectVersion

See: newCreateProjectVersion smart constructor.

Constructors

CreateProjectVersion'

Fields

kmsKeyId :: Maybe Text

The identifier for your AWS Key Management Service (AWS KMS) customer master key (CMK). You can supply the Amazon Resource Name (ARN) of your CMK, the ID of your CMK, an alias for your CMK, or an alias ARN. The key is used to encrypt training and test images copied into the service for model training. Your source images are unaffected. The key is also used to encrypt training results and manifest files written to the output Amazon S3 bucket (OutputConfig).

If you choose to use your own CMK, you need the following permissions on the CMK.

kms:CreateGrant

kms:DescribeKey

kms:GenerateDataKey

kms:Decrypt

If you don't specify a value for KmsKeyId, images copied into the service are encrypted using a key that AWS owns and manages.

tags :: Maybe (HashMap Text Text)

A set of tags (key-value pairs) that you want to attach to the model.

projectArn :: Text

The ARN of the Amazon Rekognition Custom Labels project that manages the model that you want to train.

versionName :: Text

A name for the version of the model. This value must be unique.

outputConfig :: OutputConfig

The Amazon S3 bucket location to store the results of training. The S3 bucket can be in any AWS account as long as the caller has s3:PutObject permissions on the S3 bucket.

trainingData :: TrainingData

The dataset to use for training.

testingData :: TestingData

The dataset to use for testing.

newCreateProjectVersion

Arguments

:: Text
$sel:projectArn:CreateProjectVersion'

-> Text
$sel:versionName:CreateProjectVersion'

-> OutputConfig
$sel:outputConfig:CreateProjectVersion'

-> TrainingData
$sel:trainingData:CreateProjectVersion'

-> TestingData
$sel:testingData:CreateProjectVersion'

-> CreateProjectVersion

Create a value of CreateProjectVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyId:CreateProjectVersion', createProjectVersion_kmsKeyId - The identifier for your AWS Key Management Service (AWS KMS) customer master key (CMK). You can supply the Amazon Resource Name (ARN) of your CMK, the ID of your CMK, an alias for your CMK, or an alias ARN. The key is used to encrypt training and test images copied into the service for model training. Your source images are unaffected. The key is also used to encrypt training results and manifest files written to the output Amazon S3 bucket (OutputConfig).

If you choose to use your own CMK, you need the following permissions on the CMK.

kms:CreateGrant

kms:DescribeKey

kms:GenerateDataKey

kms:Decrypt

If you don't specify a value for KmsKeyId, images copied into the service are encrypted using a key that AWS owns and manages.

$sel:tags:CreateProjectVersion', createProjectVersion_tags - A set of tags (key-value pairs) that you want to attach to the model.

$sel:projectArn:CreateProjectVersion', createProjectVersion_projectArn - The ARN of the Amazon Rekognition Custom Labels project that manages the model that you want to train.

$sel:versionName:CreateProjectVersion', createProjectVersion_versionName - A name for the version of the model. This value must be unique.

$sel:outputConfig:CreateProjectVersion', createProjectVersion_outputConfig - The Amazon S3 bucket location to store the results of training. The S3 bucket can be in any AWS account as long as the caller has s3:PutObject permissions on the S3 bucket.

$sel:trainingData:CreateProjectVersion', createProjectVersion_trainingData - The dataset to use for training.

$sel:testingData:CreateProjectVersion', createProjectVersion_testingData - The dataset to use for testing.

Request Lenses

createProjectVersion_kmsKeyId :: Lens' CreateProjectVersion (Maybe Text)

The identifier for your AWS Key Management Service (AWS KMS) customer master key (CMK). You can supply the Amazon Resource Name (ARN) of your CMK, the ID of your CMK, an alias for your CMK, or an alias ARN. The key is used to encrypt training and test images copied into the service for model training. Your source images are unaffected. The key is also used to encrypt training results and manifest files written to the output Amazon S3 bucket (OutputConfig).

If you choose to use your own CMK, you need the following permissions on the CMK.

kms:CreateGrant

kms:DescribeKey

kms:GenerateDataKey

kms:Decrypt

If you don't specify a value for KmsKeyId, images copied into the service are encrypted using a key that AWS owns and manages.

createProjectVersion_tags :: Lens' CreateProjectVersion (Maybe (HashMap Text Text))

A set of tags (key-value pairs) that you want to attach to the model.

createProjectVersion_projectArn :: Lens' CreateProjectVersion Text

The ARN of the Amazon Rekognition Custom Labels project that manages the model that you want to train.

createProjectVersion_versionName :: Lens' CreateProjectVersion Text

A name for the version of the model. This value must be unique.

createProjectVersion_outputConfig :: Lens' CreateProjectVersion OutputConfig

The Amazon S3 bucket location to store the results of training. The S3 bucket can be in any AWS account as long as the caller has s3:PutObject permissions on the S3 bucket.

createProjectVersion_trainingData :: Lens' CreateProjectVersion TrainingData

The dataset to use for training.

createProjectVersion_testingData :: Lens' CreateProjectVersion TestingData

The dataset to use for testing.

Destructuring the Response

data CreateProjectVersionResponse

See: newCreateProjectVersionResponse smart constructor.

Constructors

CreateProjectVersionResponse'

Fields

projectVersionArn :: Maybe Text

The ARN of the model version that was created. Use DescribeProjectVersion to get the current status of the training operation.

httpStatus :: Int

The response's http status code.

newCreateProjectVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateProjectVersionResponse'

-> CreateProjectVersionResponse

Create a value of CreateProjectVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectVersionArn:CreateProjectVersionResponse', createProjectVersionResponse_projectVersionArn - The ARN of the model version that was created. Use DescribeProjectVersion to get the current status of the training operation.

$sel:httpStatus:CreateProjectVersionResponse', createProjectVersionResponse_httpStatus - The response's http status code.

Response Lenses

createProjectVersionResponse_projectVersionArn :: Lens' CreateProjectVersionResponse (Maybe Text)

The ARN of the model version that was created. Use DescribeProjectVersion to get the current status of the training operation.

createProjectVersionResponse_httpStatus :: Lens' CreateProjectVersionResponse Int

The response's http status code.