Network.AWS.Rekognition.DescribeCollection

Description

Describes the specified collection. You can use DescribeCollection to get information, such as the number of faces indexed into a collection and the version of the model used by the collection for face detection.

For more information, see Describing a Collection in the Amazon Rekognition Developer Guide.

Creating a Request

data DescribeCollection

See: newDescribeCollection smart constructor.

Constructors

DescribeCollection'

Fields

collectionId :: Text

The ID of the collection to describe.

newDescribeCollection

Arguments

:: Text
$sel:collectionId:DescribeCollection'

-> DescribeCollection

Create a value of DescribeCollection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:collectionId:DescribeCollection', describeCollection_collectionId - The ID of the collection to describe.

Request Lenses

describeCollection_collectionId :: Lens' DescribeCollection Text

The ID of the collection to describe.

Destructuring the Response

data DescribeCollectionResponse

See: newDescribeCollectionResponse smart constructor.

Constructors

DescribeCollectionResponse'

Fields

creationTimestamp :: Maybe POSIX

The number of milliseconds since the Unix epoch time until the creation of the collection. The Unix epoch time is 00:00:00 Coordinated Universal Time (UTC), Thursday, 1 January 1970.

faceModelVersion :: Maybe Text

The version of the face model that's used by the collection for face detection.

For more information, see Model Versioning in the Amazon Rekognition Developer Guide.

collectionARN :: Maybe Text

The Amazon Resource Name (ARN) of the collection.

faceCount :: Maybe Natural

The number of faces that are indexed into the collection. To index faces into a collection, use IndexFaces.

httpStatus :: Int

The response's http status code.

newDescribeCollectionResponse

Arguments

:: Int
$sel:httpStatus:DescribeCollectionResponse'

-> DescribeCollectionResponse

Create a value of DescribeCollectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTimestamp:DescribeCollectionResponse', describeCollectionResponse_creationTimestamp - The number of milliseconds since the Unix epoch time until the creation of the collection. The Unix epoch time is 00:00:00 Coordinated Universal Time (UTC), Thursday, 1 January 1970.

$sel:faceModelVersion:DescribeCollectionResponse', describeCollectionResponse_faceModelVersion - The version of the face model that's used by the collection for face detection.

For more information, see Model Versioning in the Amazon Rekognition Developer Guide.

$sel:collectionARN:DescribeCollectionResponse', describeCollectionResponse_collectionARN - The Amazon Resource Name (ARN) of the collection.

$sel:faceCount:DescribeCollectionResponse', describeCollectionResponse_faceCount - The number of faces that are indexed into the collection. To index faces into a collection, use IndexFaces.

$sel:httpStatus:DescribeCollectionResponse', describeCollectionResponse_httpStatus - The response's http status code.

Response Lenses

describeCollectionResponse_creationTimestamp :: Lens' DescribeCollectionResponse (Maybe UTCTime)

The number of milliseconds since the Unix epoch time until the creation of the collection. The Unix epoch time is 00:00:00 Coordinated Universal Time (UTC), Thursday, 1 January 1970.

describeCollectionResponse_faceModelVersion :: Lens' DescribeCollectionResponse (Maybe Text)

The version of the face model that's used by the collection for face detection.

For more information, see Model Versioning in the Amazon Rekognition Developer Guide.

describeCollectionResponse_collectionARN :: Lens' DescribeCollectionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the collection.

describeCollectionResponse_faceCount :: Lens' DescribeCollectionResponse (Maybe Natural)

The number of faces that are indexed into the collection. To index faces into a collection, use IndexFaces.

describeCollectionResponse_httpStatus :: Lens' DescribeCollectionResponse Int

The response's http status code.