Network.AWS.Rekognition.DescribeStreamProcessor

Description

Provides information about a stream processor created by CreateStreamProcessor. You can get information about the input and output streams, the input parameters for the face recognition being performed, and the current status of the stream processor.

Creating a Request

data DescribeStreamProcessor

See: newDescribeStreamProcessor smart constructor.

Constructors

DescribeStreamProcessor'

Fields

name :: Text

Name of the stream processor for which you want information.

newDescribeStreamProcessor

Arguments

:: Text
$sel:name:DescribeStreamProcessor'

-> DescribeStreamProcessor

Create a value of DescribeStreamProcessor with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DescribeStreamProcessor', describeStreamProcessor_name - Name of the stream processor for which you want information.

Request Lenses

describeStreamProcessor_name :: Lens' DescribeStreamProcessor Text

Name of the stream processor for which you want information.

Destructuring the Response

data DescribeStreamProcessorResponse

See: newDescribeStreamProcessorResponse smart constructor.

Constructors

DescribeStreamProcessorResponse'

Fields

creationTimestamp :: Maybe POSIX

Date and time the stream processor was created

statusMessage :: Maybe Text

Detailed status message about the stream processor.

status :: Maybe StreamProcessorStatus

Current status of the stream processor.

roleArn :: Maybe Text

ARN of the IAM role that allows access to the stream processor.

input :: Maybe StreamProcessorInput

Kinesis video stream that provides the source streaming video.

name :: Maybe Text

Name of the stream processor.

streamProcessorArn :: Maybe Text

ARN of the stream processor.

output :: Maybe StreamProcessorOutput

Kinesis data stream to which Amazon Rekognition Video puts the analysis results.

lastUpdateTimestamp :: Maybe POSIX

The time, in Unix format, the stream processor was last updated. For example, when the stream processor moves from a running state to a failed state, or when the user starts or stops the stream processor.

settings :: Maybe StreamProcessorSettings

Face recognition input parameters that are being used by the stream processor. Includes the collection to use for face recognition and the face attributes to detect.

httpStatus :: Int

The response's http status code.

newDescribeStreamProcessorResponse

Arguments

:: Int
$sel:httpStatus:DescribeStreamProcessorResponse'

-> DescribeStreamProcessorResponse

Create a value of DescribeStreamProcessorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTimestamp:DescribeStreamProcessorResponse', describeStreamProcessorResponse_creationTimestamp - Date and time the stream processor was created

$sel:statusMessage:DescribeStreamProcessorResponse', describeStreamProcessorResponse_statusMessage - Detailed status message about the stream processor.

$sel:status:DescribeStreamProcessorResponse', describeStreamProcessorResponse_status - Current status of the stream processor.

$sel:roleArn:DescribeStreamProcessorResponse', describeStreamProcessorResponse_roleArn - ARN of the IAM role that allows access to the stream processor.

$sel:input:DescribeStreamProcessorResponse', describeStreamProcessorResponse_input - Kinesis video stream that provides the source streaming video.

$sel:name:DescribeStreamProcessor', describeStreamProcessorResponse_name - Name of the stream processor.

$sel:streamProcessorArn:DescribeStreamProcessorResponse', describeStreamProcessorResponse_streamProcessorArn - ARN of the stream processor.

$sel:output:DescribeStreamProcessorResponse', describeStreamProcessorResponse_output - Kinesis data stream to which Amazon Rekognition Video puts the analysis results.

$sel:lastUpdateTimestamp:DescribeStreamProcessorResponse', describeStreamProcessorResponse_lastUpdateTimestamp - The time, in Unix format, the stream processor was last updated. For example, when the stream processor moves from a running state to a failed state, or when the user starts or stops the stream processor.

$sel:settings:DescribeStreamProcessorResponse', describeStreamProcessorResponse_settings - Face recognition input parameters that are being used by the stream processor. Includes the collection to use for face recognition and the face attributes to detect.

$sel:httpStatus:DescribeStreamProcessorResponse', describeStreamProcessorResponse_httpStatus - The response's http status code.

Response Lenses

describeStreamProcessorResponse_creationTimestamp :: Lens' DescribeStreamProcessorResponse (Maybe UTCTime)

Date and time the stream processor was created

describeStreamProcessorResponse_statusMessage :: Lens' DescribeStreamProcessorResponse (Maybe Text)

Detailed status message about the stream processor.

describeStreamProcessorResponse_status :: Lens' DescribeStreamProcessorResponse (Maybe StreamProcessorStatus)

Current status of the stream processor.

describeStreamProcessorResponse_roleArn :: Lens' DescribeStreamProcessorResponse (Maybe Text)

ARN of the IAM role that allows access to the stream processor.

describeStreamProcessorResponse_input :: Lens' DescribeStreamProcessorResponse (Maybe StreamProcessorInput)

Kinesis video stream that provides the source streaming video.

describeStreamProcessorResponse_name :: Lens' DescribeStreamProcessorResponse (Maybe Text)

Name of the stream processor.

describeStreamProcessorResponse_streamProcessorArn :: Lens' DescribeStreamProcessorResponse (Maybe Text)

ARN of the stream processor.

describeStreamProcessorResponse_output :: Lens' DescribeStreamProcessorResponse (Maybe StreamProcessorOutput)

Kinesis data stream to which Amazon Rekognition Video puts the analysis results.

describeStreamProcessorResponse_lastUpdateTimestamp :: Lens' DescribeStreamProcessorResponse (Maybe UTCTime)

The time, in Unix format, the stream processor was last updated. For example, when the stream processor moves from a running state to a failed state, or when the user starts or stops the stream processor.

describeStreamProcessorResponse_settings :: Lens' DescribeStreamProcessorResponse (Maybe StreamProcessorSettings)

Face recognition input parameters that are being used by the stream processor. Includes the collection to use for face recognition and the face attributes to detect.

describeStreamProcessorResponse_httpStatus :: Lens' DescribeStreamProcessorResponse Int

The response's http status code.