Network.AWS.Rekognition.DetectCustomLabels

Description

Detects custom labels in a supplied image by using an Amazon Rekognition Custom Labels model.

You specify which version of a model version to use by using the ProjectVersionArn input parameter.

You pass the input image as base64-encoded image bytes or as a reference to an image in an Amazon S3 bucket. If you use the AWS CLI to call Amazon Rekognition operations, passing image bytes is not supported. The image must be either a PNG or JPEG formatted file.

For each object that the model version detects on an image, the API returns a (CustomLabel) object in an array (CustomLabels). Each CustomLabel object provides the label name (Name), the level of confidence that the image contains the object (Confidence), and object location information, if it exists, for the label on the image (Geometry).

To filter labels that are returned, specify a value for MinConfidence. DetectCustomLabelsLabels only returns labels with a confidence that's higher than the specified value. The value of MinConfidence maps to the assumed threshold values created during training. For more information, see Assumed threshold in the Amazon Rekognition Custom Labels Developer Guide. Amazon Rekognition Custom Labels metrics expresses an assumed threshold as a floating point value between 0-1. The range of MinConfidence normalizes the threshold value to a percentage value (0-100). Confidence responses from DetectCustomLabels are also returned as a percentage. You can use MinConfidence to change the precision and recall or your model. For more information, see Analyzing an image in the Amazon Rekognition Custom Labels Developer Guide.

If you don't specify a value for MinConfidence, DetectCustomLabels returns labels based on the assumed threshold of each label.

This is a stateless API operation. That is, the operation does not persist any data.

This operation requires permissions to perform the rekognition:DetectCustomLabels action.

For more information, see Analyzing an image in the Amazon Rekognition Custom Labels Developer Guide.

Creating a Request

data DetectCustomLabels

See: newDetectCustomLabels smart constructor.

Constructors

DetectCustomLabels'

Fields

maxResults :: Maybe Natural

Maximum number of results you want the service to return in the response. The service returns the specified number of highest confidence labels ranked from highest confidence to lowest.

minConfidence :: Maybe Double

Specifies the minimum confidence level for the labels to return. DetectCustomLabels doesn't return any labels with a confidence value that's lower than this specified value. If you specify a value of 0, DetectCustomLabels returns all labels, regardless of the assumed threshold applied to each label. If you don't specify a value for MinConfidence, DetectCustomLabels returns labels based on the assumed threshold of each label.

projectVersionArn :: Text

The ARN of the model version that you want to use.

image :: Image

newDetectCustomLabels

Arguments

:: Text
$sel:projectVersionArn:DetectCustomLabels'

-> Image
$sel:image:DetectCustomLabels'

-> DetectCustomLabels

Create a value of DetectCustomLabels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxResults:DetectCustomLabels', detectCustomLabels_maxResults - Maximum number of results you want the service to return in the response. The service returns the specified number of highest confidence labels ranked from highest confidence to lowest.

$sel:minConfidence:DetectCustomLabels', detectCustomLabels_minConfidence - Specifies the minimum confidence level for the labels to return. DetectCustomLabels doesn't return any labels with a confidence value that's lower than this specified value. If you specify a value of 0, DetectCustomLabels returns all labels, regardless of the assumed threshold applied to each label. If you don't specify a value for MinConfidence, DetectCustomLabels returns labels based on the assumed threshold of each label.

$sel:projectVersionArn:DetectCustomLabels', detectCustomLabels_projectVersionArn - The ARN of the model version that you want to use.

$sel:image:DetectCustomLabels', detectCustomLabels_image - Undocumented member.

Request Lenses

detectCustomLabels_maxResults :: Lens' DetectCustomLabels (Maybe Natural)

Maximum number of results you want the service to return in the response. The service returns the specified number of highest confidence labels ranked from highest confidence to lowest.

detectCustomLabels_minConfidence :: Lens' DetectCustomLabels (Maybe Double)

Specifies the minimum confidence level for the labels to return. DetectCustomLabels doesn't return any labels with a confidence value that's lower than this specified value. If you specify a value of 0, DetectCustomLabels returns all labels, regardless of the assumed threshold applied to each label. If you don't specify a value for MinConfidence, DetectCustomLabels returns labels based on the assumed threshold of each label.

detectCustomLabels_projectVersionArn :: Lens' DetectCustomLabels Text

The ARN of the model version that you want to use.

detectCustomLabels_image :: Lens' DetectCustomLabels Image

Undocumented member.

Destructuring the Response

data DetectCustomLabelsResponse

See: newDetectCustomLabelsResponse smart constructor.

Constructors

DetectCustomLabelsResponse'

Fields

customLabels :: Maybe [CustomLabel]

An array of custom labels detected in the input image.

httpStatus :: Int

The response's http status code.

newDetectCustomLabelsResponse

Arguments

:: Int
$sel:httpStatus:DetectCustomLabelsResponse'

-> DetectCustomLabelsResponse

Create a value of DetectCustomLabelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customLabels:DetectCustomLabelsResponse', detectCustomLabelsResponse_customLabels - An array of custom labels detected in the input image.

$sel:httpStatus:DetectCustomLabelsResponse', detectCustomLabelsResponse_httpStatus - The response's http status code.

Response Lenses

detectCustomLabelsResponse_customLabels :: Lens' DetectCustomLabelsResponse (Maybe [CustomLabel])

An array of custom labels detected in the input image.

detectCustomLabelsResponse_httpStatus :: Lens' DetectCustomLabelsResponse Int

The response's http status code.