Network.AWS.Rekognition.DetectModerationLabels

Description

Detects unsafe content in a specified JPEG or PNG format image. Use DetectModerationLabels to moderate images depending on your requirements. For example, you might want to filter images that contain nudity, but not images containing suggestive content.

To filter images, use the labels returned by DetectModerationLabels to determine which types of content are appropriate.

For information about moderation labels, see Detecting Unsafe Content in the Amazon Rekognition Developer Guide.

You pass the input image either as base64-encoded image bytes or as a reference to an image in an Amazon S3 bucket. If you use the AWS CLI to call Amazon Rekognition operations, passing image bytes is not supported. The image must be either a PNG or JPEG formatted file.

Creating a Request

data DetectModerationLabels

See: newDetectModerationLabels smart constructor.

Constructors

DetectModerationLabels'

Fields

humanLoopConfig :: Maybe HumanLoopConfig

Sets up the configuration for human evaluation, including the FlowDefinition the image will be sent to.

minConfidence :: Maybe Double

Specifies the minimum confidence level for the labels to return. Amazon Rekognition doesn't return any labels with a confidence level lower than this specified value.

If you don't specify MinConfidence, the operation returns labels with confidence values greater than or equal to 50 percent.

image :: Image

The input image as base64-encoded bytes or an S3 object. If you use the AWS CLI to call Amazon Rekognition operations, passing base64-encoded image bytes is not supported.

If you are using an AWS SDK to call Amazon Rekognition, you might not need to base64-encode image bytes passed using the Bytes field. For more information, see Images in the Amazon Rekognition developer guide.

newDetectModerationLabels

Arguments

:: Image
$sel:image:DetectModerationLabels'

-> DetectModerationLabels

Create a value of DetectModerationLabels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:humanLoopConfig:DetectModerationLabels', detectModerationLabels_humanLoopConfig - Sets up the configuration for human evaluation, including the FlowDefinition the image will be sent to.

$sel:minConfidence:DetectModerationLabels', detectModerationLabels_minConfidence - Specifies the minimum confidence level for the labels to return. Amazon Rekognition doesn't return any labels with a confidence level lower than this specified value.

If you don't specify MinConfidence, the operation returns labels with confidence values greater than or equal to 50 percent.

$sel:image:DetectModerationLabels', detectModerationLabels_image - The input image as base64-encoded bytes or an S3 object. If you use the AWS CLI to call Amazon Rekognition operations, passing base64-encoded image bytes is not supported.

If you are using an AWS SDK to call Amazon Rekognition, you might not need to base64-encode image bytes passed using the Bytes field. For more information, see Images in the Amazon Rekognition developer guide.

Request Lenses

detectModerationLabels_humanLoopConfig :: Lens' DetectModerationLabels (Maybe HumanLoopConfig)

Sets up the configuration for human evaluation, including the FlowDefinition the image will be sent to.

detectModerationLabels_minConfidence :: Lens' DetectModerationLabels (Maybe Double)

Specifies the minimum confidence level for the labels to return. Amazon Rekognition doesn't return any labels with a confidence level lower than this specified value.

If you don't specify MinConfidence, the operation returns labels with confidence values greater than or equal to 50 percent.

detectModerationLabels_image :: Lens' DetectModerationLabels Image

The input image as base64-encoded bytes or an S3 object. If you use the AWS CLI to call Amazon Rekognition operations, passing base64-encoded image bytes is not supported.

If you are using an AWS SDK to call Amazon Rekognition, you might not need to base64-encode image bytes passed using the Bytes field. For more information, see Images in the Amazon Rekognition developer guide.

Destructuring the Response

data DetectModerationLabelsResponse

See: newDetectModerationLabelsResponse smart constructor.

Constructors

DetectModerationLabelsResponse'

Fields

humanLoopActivationOutput :: Maybe HumanLoopActivationOutput

Shows the results of the human in the loop evaluation.

moderationModelVersion :: Maybe Text

Version number of the moderation detection model that was used to detect unsafe content.

moderationLabels :: Maybe [ModerationLabel]

Array of detected Moderation labels and the time, in milliseconds from the start of the video, they were detected.

httpStatus :: Int

The response's http status code.

newDetectModerationLabelsResponse

Arguments

:: Int
$sel:httpStatus:DetectModerationLabelsResponse'

-> DetectModerationLabelsResponse

Create a value of DetectModerationLabelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:humanLoopActivationOutput:DetectModerationLabelsResponse', detectModerationLabelsResponse_humanLoopActivationOutput - Shows the results of the human in the loop evaluation.

$sel:moderationModelVersion:DetectModerationLabelsResponse', detectModerationLabelsResponse_moderationModelVersion - Version number of the moderation detection model that was used to detect unsafe content.

$sel:moderationLabels:DetectModerationLabelsResponse', detectModerationLabelsResponse_moderationLabels - Array of detected Moderation labels and the time, in milliseconds from the start of the video, they were detected.

$sel:httpStatus:DetectModerationLabelsResponse', detectModerationLabelsResponse_httpStatus - The response's http status code.

Response Lenses

detectModerationLabelsResponse_humanLoopActivationOutput :: Lens' DetectModerationLabelsResponse (Maybe HumanLoopActivationOutput)

Shows the results of the human in the loop evaluation.

detectModerationLabelsResponse_moderationModelVersion :: Lens' DetectModerationLabelsResponse (Maybe Text)

Version number of the moderation detection model that was used to detect unsafe content.

detectModerationLabelsResponse_moderationLabels :: Lens' DetectModerationLabelsResponse (Maybe [ModerationLabel])

Array of detected Moderation labels and the time, in milliseconds from the start of the video, they were detected.

detectModerationLabelsResponse_httpStatus :: Lens' DetectModerationLabelsResponse Int

The response's http status code.