Network.AWS.Rekognition.GetCelebrityInfo

Description

Gets the name and additional information about a celebrity based on their Amazon Rekognition ID. The additional information is returned as an array of URLs. If there is no additional information about the celebrity, this list is empty.

For more information, see Recognizing Celebrities in an Image in the Amazon Rekognition Developer Guide.

This operation requires permissions to perform the rekognition:GetCelebrityInfo action.

Creating a Request

data GetCelebrityInfo

See: newGetCelebrityInfo smart constructor.

Constructors

GetCelebrityInfo'

Fields

id :: Text

The ID for the celebrity. You get the celebrity ID from a call to the RecognizeCelebrities operation, which recognizes celebrities in an image.

newGetCelebrityInfo

Arguments

:: Text
$sel:id:GetCelebrityInfo'

-> GetCelebrityInfo

Create a value of GetCelebrityInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:GetCelebrityInfo', getCelebrityInfo_id - The ID for the celebrity. You get the celebrity ID from a call to the RecognizeCelebrities operation, which recognizes celebrities in an image.

Request Lenses

getCelebrityInfo_id :: Lens' GetCelebrityInfo Text

The ID for the celebrity. You get the celebrity ID from a call to the RecognizeCelebrities operation, which recognizes celebrities in an image.

Destructuring the Response

data GetCelebrityInfoResponse

See: newGetCelebrityInfoResponse smart constructor.

Constructors

GetCelebrityInfoResponse'

Fields

urls :: Maybe [Text]

An array of URLs pointing to additional celebrity information.

knownGender :: Maybe KnownGender

Retrieves the known gender for the celebrity.

name :: Maybe Text

The name of the celebrity.

httpStatus :: Int

The response's http status code.

newGetCelebrityInfoResponse

Arguments

:: Int
$sel:httpStatus:GetCelebrityInfoResponse'

-> GetCelebrityInfoResponse

Create a value of GetCelebrityInfoResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:urls:GetCelebrityInfoResponse', getCelebrityInfoResponse_urls - An array of URLs pointing to additional celebrity information.

$sel:knownGender:GetCelebrityInfoResponse', getCelebrityInfoResponse_knownGender - Retrieves the known gender for the celebrity.

$sel:name:GetCelebrityInfoResponse', getCelebrityInfoResponse_name - The name of the celebrity.

$sel:httpStatus:GetCelebrityInfoResponse', getCelebrityInfoResponse_httpStatus - The response's http status code.

Response Lenses

getCelebrityInfoResponse_urls :: Lens' GetCelebrityInfoResponse (Maybe [Text])

An array of URLs pointing to additional celebrity information.

getCelebrityInfoResponse_knownGender :: Lens' GetCelebrityInfoResponse (Maybe KnownGender)

Retrieves the known gender for the celebrity.

getCelebrityInfoResponse_name :: Lens' GetCelebrityInfoResponse (Maybe Text)

The name of the celebrity.

getCelebrityInfoResponse_httpStatus :: Lens' GetCelebrityInfoResponse Int

The response's http status code.