Network.AWS.Rekognition.GetContentModeration

Description

Gets the inappropriate, unwanted, or offensive content analysis results for a Amazon Rekognition Video analysis started by StartContentModeration. For a list of moderation labels in Amazon Rekognition, see Using the image and video moderation APIs.

Amazon Rekognition Video inappropriate or offensive content detection in a stored video is an asynchronous operation. You start analysis by calling StartContentModeration which returns a job identifier (JobId). When analysis finishes, Amazon Rekognition Video publishes a completion status to the Amazon Simple Notification Service topic registered in the initial call to StartContentModeration. To get the results of the content analysis, first check that the status value published to the Amazon SNS topic is SUCCEEDED. If so, call GetContentModeration and pass the job identifier (JobId) from the initial call to StartContentModeration.

For more information, see Working with Stored Videos in the Amazon Rekognition Devlopers Guide.

GetContentModeration returns detected inappropriate, unwanted, or offensive content moderation labels, and the time they are detected, in an array, ModerationLabels, of ContentModerationDetection objects.

By default, the moderated labels are returned sorted by time, in milliseconds from the start of the video. You can also sort them by moderated label by specifying NAME for the SortBy input parameter.

Since video analysis can return a large number of results, use the MaxResults parameter to limit the number of labels returned in a single call to GetContentModeration. If there are more results than specified in MaxResults, the value of NextToken in the operation response contains a pagination token for getting the next set of results. To get the next page of results, call GetContentModeration and populate the NextToken request parameter with the value of NextToken returned from the previous call to GetContentModeration.

For more information, see Content moderation in the Amazon Rekognition Developer Guide.

Creating a Request

data GetContentModeration

See: newGetContentModeration smart constructor.

Constructors

GetContentModeration'

Fields

nextToken :: Maybe Text

If the previous response was incomplete (because there is more data to retrieve), Amazon Rekognition returns a pagination token in the response. You can use this pagination token to retrieve the next set of content moderation labels.

maxResults :: Maybe Natural

Maximum number of results to return per paginated call. The largest value you can specify is 1000. If you specify a value greater than 1000, a maximum of 1000 results is returned. The default value is 1000.

sortBy :: Maybe ContentModerationSortBy

Sort to use for elements in the ModerationLabelDetections array. Use TIMESTAMP to sort array elements by the time labels are detected. Use NAME to alphabetically group elements for a label together. Within each label group, the array element are sorted by detection confidence. The default sort is by TIMESTAMP.

jobId :: Text

The identifier for the inappropriate, unwanted, or offensive content moderation job. Use JobId to identify the job in a subsequent call to GetContentModeration.

newGetContentModeration

Arguments

:: Text
$sel:jobId:GetContentModeration'

-> GetContentModeration

Create a value of GetContentModeration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetContentModeration', getContentModeration_nextToken - If the previous response was incomplete (because there is more data to retrieve), Amazon Rekognition returns a pagination token in the response. You can use this pagination token to retrieve the next set of content moderation labels.

$sel:maxResults:GetContentModeration', getContentModeration_maxResults - Maximum number of results to return per paginated call. The largest value you can specify is 1000. If you specify a value greater than 1000, a maximum of 1000 results is returned. The default value is 1000.

$sel:sortBy:GetContentModeration', getContentModeration_sortBy - Sort to use for elements in the ModerationLabelDetections array. Use TIMESTAMP to sort array elements by the time labels are detected. Use NAME to alphabetically group elements for a label together. Within each label group, the array element are sorted by detection confidence. The default sort is by TIMESTAMP.

$sel:jobId:GetContentModeration', getContentModeration_jobId - The identifier for the inappropriate, unwanted, or offensive content moderation job. Use JobId to identify the job in a subsequent call to GetContentModeration.

Request Lenses

getContentModeration_nextToken :: Lens' GetContentModeration (Maybe Text)

If the previous response was incomplete (because there is more data to retrieve), Amazon Rekognition returns a pagination token in the response. You can use this pagination token to retrieve the next set of content moderation labels.

getContentModeration_maxResults :: Lens' GetContentModeration (Maybe Natural)

Maximum number of results to return per paginated call. The largest value you can specify is 1000. If you specify a value greater than 1000, a maximum of 1000 results is returned. The default value is 1000.

getContentModeration_sortBy :: Lens' GetContentModeration (Maybe ContentModerationSortBy)

Sort to use for elements in the ModerationLabelDetections array. Use TIMESTAMP to sort array elements by the time labels are detected. Use NAME to alphabetically group elements for a label together. Within each label group, the array element are sorted by detection confidence. The default sort is by TIMESTAMP.

getContentModeration_jobId :: Lens' GetContentModeration Text

The identifier for the inappropriate, unwanted, or offensive content moderation job. Use JobId to identify the job in a subsequent call to GetContentModeration.

Destructuring the Response

data GetContentModerationResponse

See: newGetContentModerationResponse smart constructor.

Constructors

GetContentModerationResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, Amazon Rekognition Video returns this token that you can use in the subsequent request to retrieve the next set of content moderation labels.

videoMetadata :: Maybe VideoMetadata

Information about a video that Amazon Rekognition analyzed. Videometadata is returned in every page of paginated responses from GetContentModeration.

statusMessage :: Maybe Text

If the job fails, StatusMessage provides a descriptive error message.

jobStatus :: Maybe VideoJobStatus

The current status of the content moderation analysis job.

moderationModelVersion :: Maybe Text

Version number of the moderation detection model that was used to detect inappropriate, unwanted, or offensive content.

moderationLabels :: Maybe [ContentModerationDetection]

The detected inappropriate, unwanted, or offensive content moderation labels and the time(s) they were detected.

httpStatus :: Int

The response's http status code.

newGetContentModerationResponse

Arguments

:: Int
$sel:httpStatus:GetContentModerationResponse'

-> GetContentModerationResponse

Create a value of GetContentModerationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetContentModeration', getContentModerationResponse_nextToken - If the response is truncated, Amazon Rekognition Video returns this token that you can use in the subsequent request to retrieve the next set of content moderation labels.

$sel:videoMetadata:GetContentModerationResponse', getContentModerationResponse_videoMetadata - Information about a video that Amazon Rekognition analyzed. Videometadata is returned in every page of paginated responses from GetContentModeration.

$sel:statusMessage:GetContentModerationResponse', getContentModerationResponse_statusMessage - If the job fails, StatusMessage provides a descriptive error message.

$sel:jobStatus:GetContentModerationResponse', getContentModerationResponse_jobStatus - The current status of the content moderation analysis job.

$sel:moderationModelVersion:GetContentModerationResponse', getContentModerationResponse_moderationModelVersion - Version number of the moderation detection model that was used to detect inappropriate, unwanted, or offensive content.

$sel:moderationLabels:GetContentModerationResponse', getContentModerationResponse_moderationLabels - The detected inappropriate, unwanted, or offensive content moderation labels and the time(s) they were detected.

$sel:httpStatus:GetContentModerationResponse', getContentModerationResponse_httpStatus - The response's http status code.

Response Lenses

getContentModerationResponse_nextToken :: Lens' GetContentModerationResponse (Maybe Text)

If the response is truncated, Amazon Rekognition Video returns this token that you can use in the subsequent request to retrieve the next set of content moderation labels.

getContentModerationResponse_videoMetadata :: Lens' GetContentModerationResponse (Maybe VideoMetadata)

Information about a video that Amazon Rekognition analyzed. Videometadata is returned in every page of paginated responses from GetContentModeration.

getContentModerationResponse_statusMessage :: Lens' GetContentModerationResponse (Maybe Text)

If the job fails, StatusMessage provides a descriptive error message.

getContentModerationResponse_jobStatus :: Lens' GetContentModerationResponse (Maybe VideoJobStatus)

The current status of the content moderation analysis job.

getContentModerationResponse_moderationModelVersion :: Lens' GetContentModerationResponse (Maybe Text)

Version number of the moderation detection model that was used to detect inappropriate, unwanted, or offensive content.

getContentModerationResponse_moderationLabels :: Lens' GetContentModerationResponse (Maybe [ContentModerationDetection])

The detected inappropriate, unwanted, or offensive content moderation labels and the time(s) they were detected.

getContentModerationResponse_httpStatus :: Lens' GetContentModerationResponse Int

The response's http status code.