Network.AWS.Rekognition.GetLabelDetection

Description

Gets the label detection results of a Amazon Rekognition Video analysis started by StartLabelDetection.

The label detection operation is started by a call to StartLabelDetection which returns a job identifier (JobId). When the label detection operation finishes, Amazon Rekognition publishes a completion status to the Amazon Simple Notification Service topic registered in the initial call to StartlabelDetection. To get the results of the label detection operation, first check that the status value published to the Amazon SNS topic is SUCCEEDED. If so, call GetLabelDetection and pass the job identifier (JobId) from the initial call to StartLabelDetection.

GetLabelDetection returns an array of detected labels (Labels) sorted by the time the labels were detected. You can also sort by the label name by specifying NAME for the SortBy input parameter.

The labels returned include the label name, the percentage confidence in the accuracy of the detected label, and the time the label was detected in the video.

The returned labels also include bounding box information for common objects, a hierarchical taxonomy of detected labels, and the version of the label model used for detection.

Use MaxResults parameter to limit the number of labels returned. If there are more results than specified in MaxResults, the value of NextToken in the operation response contains a pagination token for getting the next set of results. To get the next page of results, call GetlabelDetection and populate the NextToken request parameter with the token value returned from the previous call to GetLabelDetection.

Creating a Request

data GetLabelDetection

See: newGetLabelDetection smart constructor.

Constructors

GetLabelDetection'

Fields

nextToken :: Maybe Text

If the previous response was incomplete (because there are more labels to retrieve), Amazon Rekognition Video returns a pagination token in the response. You can use this pagination token to retrieve the next set of labels.

maxResults :: Maybe Natural

Maximum number of results to return per paginated call. The largest value you can specify is 1000. If you specify a value greater than 1000, a maximum of 1000 results is returned. The default value is 1000.

sortBy :: Maybe LabelDetectionSortBy

Sort to use for elements in the Labels array. Use TIMESTAMP to sort array elements by the time labels are detected. Use NAME to alphabetically group elements for a label together. Within each label group, the array element are sorted by detection confidence. The default sort is by TIMESTAMP.

jobId :: Text

Job identifier for the label detection operation for which you want results returned. You get the job identifer from an initial call to StartlabelDetection.

newGetLabelDetection

Arguments

:: Text
$sel:jobId:GetLabelDetection'

-> GetLabelDetection

Create a value of GetLabelDetection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetLabelDetection', getLabelDetection_nextToken - If the previous response was incomplete (because there are more labels to retrieve), Amazon Rekognition Video returns a pagination token in the response. You can use this pagination token to retrieve the next set of labels.

$sel:maxResults:GetLabelDetection', getLabelDetection_maxResults - Maximum number of results to return per paginated call. The largest value you can specify is 1000. If you specify a value greater than 1000, a maximum of 1000 results is returned. The default value is 1000.

$sel:sortBy:GetLabelDetection', getLabelDetection_sortBy - Sort to use for elements in the Labels array. Use TIMESTAMP to sort array elements by the time labels are detected. Use NAME to alphabetically group elements for a label together. Within each label group, the array element are sorted by detection confidence. The default sort is by TIMESTAMP.

$sel:jobId:GetLabelDetection', getLabelDetection_jobId - Job identifier for the label detection operation for which you want results returned. You get the job identifer from an initial call to StartlabelDetection.

Request Lenses

getLabelDetection_nextToken :: Lens' GetLabelDetection (Maybe Text)

If the previous response was incomplete (because there are more labels to retrieve), Amazon Rekognition Video returns a pagination token in the response. You can use this pagination token to retrieve the next set of labels.

getLabelDetection_maxResults :: Lens' GetLabelDetection (Maybe Natural)

Maximum number of results to return per paginated call. The largest value you can specify is 1000. If you specify a value greater than 1000, a maximum of 1000 results is returned. The default value is 1000.

getLabelDetection_sortBy :: Lens' GetLabelDetection (Maybe LabelDetectionSortBy)

Sort to use for elements in the Labels array. Use TIMESTAMP to sort array elements by the time labels are detected. Use NAME to alphabetically group elements for a label together. Within each label group, the array element are sorted by detection confidence. The default sort is by TIMESTAMP.

getLabelDetection_jobId :: Lens' GetLabelDetection Text

Job identifier for the label detection operation for which you want results returned. You get the job identifer from an initial call to StartlabelDetection.

Destructuring the Response

data GetLabelDetectionResponse

See: newGetLabelDetectionResponse smart constructor.

Constructors

GetLabelDetectionResponse'

Fields

statusMessage :: Maybe Text

If the job fails, StatusMessage provides a descriptive error message.

videoMetadata :: Maybe VideoMetadata

Information about a video that Amazon Rekognition Video analyzed. Videometadata is returned in every page of paginated responses from a Amazon Rekognition video operation.

nextToken :: Maybe Text

If the response is truncated, Amazon Rekognition Video returns this token that you can use in the subsequent request to retrieve the next set of labels.

labelModelVersion :: Maybe Text

Version number of the label detection model that was used to detect labels.

jobStatus :: Maybe VideoJobStatus

The current status of the label detection job.

labels :: Maybe [LabelDetection]

An array of labels detected in the video. Each element contains the detected label and the time, in milliseconds from the start of the video, that the label was detected.

httpStatus :: Int

The response's http status code.

newGetLabelDetectionResponse

Arguments

:: Int
$sel:httpStatus:GetLabelDetectionResponse'

-> GetLabelDetectionResponse

Create a value of GetLabelDetectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusMessage:GetLabelDetectionResponse', getLabelDetectionResponse_statusMessage - If the job fails, StatusMessage provides a descriptive error message.

$sel:videoMetadata:GetLabelDetectionResponse', getLabelDetectionResponse_videoMetadata - Information about a video that Amazon Rekognition Video analyzed. Videometadata is returned in every page of paginated responses from a Amazon Rekognition video operation.

$sel:nextToken:GetLabelDetection', getLabelDetectionResponse_nextToken - If the response is truncated, Amazon Rekognition Video returns this token that you can use in the subsequent request to retrieve the next set of labels.

$sel:labelModelVersion:GetLabelDetectionResponse', getLabelDetectionResponse_labelModelVersion - Version number of the label detection model that was used to detect labels.

$sel:jobStatus:GetLabelDetectionResponse', getLabelDetectionResponse_jobStatus - The current status of the label detection job.

$sel:labels:GetLabelDetectionResponse', getLabelDetectionResponse_labels - An array of labels detected in the video. Each element contains the detected label and the time, in milliseconds from the start of the video, that the label was detected.

$sel:httpStatus:GetLabelDetectionResponse', getLabelDetectionResponse_httpStatus - The response's http status code.

Response Lenses

getLabelDetectionResponse_statusMessage :: Lens' GetLabelDetectionResponse (Maybe Text)

If the job fails, StatusMessage provides a descriptive error message.

getLabelDetectionResponse_videoMetadata :: Lens' GetLabelDetectionResponse (Maybe VideoMetadata)

Information about a video that Amazon Rekognition Video analyzed. Videometadata is returned in every page of paginated responses from a Amazon Rekognition video operation.

getLabelDetectionResponse_nextToken :: Lens' GetLabelDetectionResponse (Maybe Text)

If the response is truncated, Amazon Rekognition Video returns this token that you can use in the subsequent request to retrieve the next set of labels.

getLabelDetectionResponse_labelModelVersion :: Lens' GetLabelDetectionResponse (Maybe Text)

Version number of the label detection model that was used to detect labels.

getLabelDetectionResponse_jobStatus :: Lens' GetLabelDetectionResponse (Maybe VideoJobStatus)

The current status of the label detection job.

getLabelDetectionResponse_labels :: Lens' GetLabelDetectionResponse (Maybe [LabelDetection])

An array of labels detected in the video. Each element contains the detected label and the time, in milliseconds from the start of the video, that the label was detected.

getLabelDetectionResponse_httpStatus :: Lens' GetLabelDetectionResponse Int

The response's http status code.