Network.AWS.Rekognition.GetPersonTracking

Description

Gets the path tracking results of a Amazon Rekognition Video analysis started by StartPersonTracking.

The person path tracking operation is started by a call to StartPersonTracking which returns a job identifier (JobId). When the operation finishes, Amazon Rekognition Video publishes a completion status to the Amazon Simple Notification Service topic registered in the initial call to StartPersonTracking.

To get the results of the person path tracking operation, first check that the status value published to the Amazon SNS topic is SUCCEEDED. If so, call GetPersonTracking and pass the job identifier (JobId) from the initial call to StartPersonTracking.

GetPersonTracking returns an array, Persons, of tracked persons and the time(s) their paths were tracked in the video.

GetPersonTracking only returns the default facial attributes (BoundingBox, Confidence, Landmarks, Pose, and Quality). The other facial attributes listed in the Face object of the following response syntax are not returned.

For more information, see FaceDetail in the Amazon Rekognition Developer Guide.

By default, the array is sorted by the time(s) a person's path is tracked in the video. You can sort by tracked persons by specifying INDEX for the SortBy input parameter.

Use the MaxResults parameter to limit the number of items returned. If there are more results than specified in MaxResults, the value of NextToken in the operation response contains a pagination token for getting the next set of results. To get the next page of results, call GetPersonTracking and populate the NextToken request parameter with the token value returned from the previous call to GetPersonTracking.

Creating a Request

data GetPersonTracking

See: newGetPersonTracking smart constructor.

Constructors

GetPersonTracking'

Fields

nextToken :: Maybe Text

If the previous response was incomplete (because there are more persons to retrieve), Amazon Rekognition Video returns a pagination token in the response. You can use this pagination token to retrieve the next set of persons.

maxResults :: Maybe Natural

Maximum number of results to return per paginated call. The largest value you can specify is 1000. If you specify a value greater than 1000, a maximum of 1000 results is returned. The default value is 1000.

sortBy :: Maybe PersonTrackingSortBy

Sort to use for elements in the Persons array. Use TIMESTAMP to sort array elements by the time persons are detected. Use INDEX to sort by the tracked persons. If you sort by INDEX, the array elements for each person are sorted by detection confidence. The default sort is by TIMESTAMP.

jobId :: Text

The identifier for a job that tracks persons in a video. You get the JobId from a call to StartPersonTracking.

newGetPersonTracking

Arguments

:: Text
$sel:jobId:GetPersonTracking'

-> GetPersonTracking

Create a value of GetPersonTracking with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetPersonTracking', getPersonTracking_nextToken - If the previous response was incomplete (because there are more persons to retrieve), Amazon Rekognition Video returns a pagination token in the response. You can use this pagination token to retrieve the next set of persons.

$sel:maxResults:GetPersonTracking', getPersonTracking_maxResults - Maximum number of results to return per paginated call. The largest value you can specify is 1000. If you specify a value greater than 1000, a maximum of 1000 results is returned. The default value is 1000.

$sel:sortBy:GetPersonTracking', getPersonTracking_sortBy - Sort to use for elements in the Persons array. Use TIMESTAMP to sort array elements by the time persons are detected. Use INDEX to sort by the tracked persons. If you sort by INDEX, the array elements for each person are sorted by detection confidence. The default sort is by TIMESTAMP.

$sel:jobId:GetPersonTracking', getPersonTracking_jobId - The identifier for a job that tracks persons in a video. You get the JobId from a call to StartPersonTracking.

Request Lenses

getPersonTracking_nextToken :: Lens' GetPersonTracking (Maybe Text)

If the previous response was incomplete (because there are more persons to retrieve), Amazon Rekognition Video returns a pagination token in the response. You can use this pagination token to retrieve the next set of persons.

getPersonTracking_maxResults :: Lens' GetPersonTracking (Maybe Natural)

Maximum number of results to return per paginated call. The largest value you can specify is 1000. If you specify a value greater than 1000, a maximum of 1000 results is returned. The default value is 1000.

getPersonTracking_sortBy :: Lens' GetPersonTracking (Maybe PersonTrackingSortBy)

Sort to use for elements in the Persons array. Use TIMESTAMP to sort array elements by the time persons are detected. Use INDEX to sort by the tracked persons. If you sort by INDEX, the array elements for each person are sorted by detection confidence. The default sort is by TIMESTAMP.

getPersonTracking_jobId :: Lens' GetPersonTracking Text

The identifier for a job that tracks persons in a video. You get the JobId from a call to StartPersonTracking.

Destructuring the Response

data GetPersonTrackingResponse

See: newGetPersonTrackingResponse smart constructor.

Constructors

GetPersonTrackingResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, Amazon Rekognition Video returns this token that you can use in the subsequent request to retrieve the next set of persons.

videoMetadata :: Maybe VideoMetadata

Information about a video that Amazon Rekognition Video analyzed. Videometadata is returned in every page of paginated responses from a Amazon Rekognition Video operation.

statusMessage :: Maybe Text

If the job fails, StatusMessage provides a descriptive error message.

jobStatus :: Maybe VideoJobStatus

The current status of the person tracking job.

persons :: Maybe [PersonDetection]

An array of the persons detected in the video and the time(s) their path was tracked throughout the video. An array element will exist for each time a person's path is tracked.

httpStatus :: Int

The response's http status code.

newGetPersonTrackingResponse

Arguments

:: Int
$sel:httpStatus:GetPersonTrackingResponse'

-> GetPersonTrackingResponse

Create a value of GetPersonTrackingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetPersonTracking', getPersonTrackingResponse_nextToken - If the response is truncated, Amazon Rekognition Video returns this token that you can use in the subsequent request to retrieve the next set of persons.

$sel:videoMetadata:GetPersonTrackingResponse', getPersonTrackingResponse_videoMetadata - Information about a video that Amazon Rekognition Video analyzed. Videometadata is returned in every page of paginated responses from a Amazon Rekognition Video operation.

$sel:statusMessage:GetPersonTrackingResponse', getPersonTrackingResponse_statusMessage - If the job fails, StatusMessage provides a descriptive error message.

$sel:jobStatus:GetPersonTrackingResponse', getPersonTrackingResponse_jobStatus - The current status of the person tracking job.

$sel:persons:GetPersonTrackingResponse', getPersonTrackingResponse_persons - An array of the persons detected in the video and the time(s) their path was tracked throughout the video. An array element will exist for each time a person's path is tracked.

$sel:httpStatus:GetPersonTrackingResponse', getPersonTrackingResponse_httpStatus - The response's http status code.

Response Lenses

getPersonTrackingResponse_nextToken :: Lens' GetPersonTrackingResponse (Maybe Text)

If the response is truncated, Amazon Rekognition Video returns this token that you can use in the subsequent request to retrieve the next set of persons.

getPersonTrackingResponse_videoMetadata :: Lens' GetPersonTrackingResponse (Maybe VideoMetadata)

Information about a video that Amazon Rekognition Video analyzed. Videometadata is returned in every page of paginated responses from a Amazon Rekognition Video operation.

getPersonTrackingResponse_statusMessage :: Lens' GetPersonTrackingResponse (Maybe Text)

If the job fails, StatusMessage provides a descriptive error message.

getPersonTrackingResponse_jobStatus :: Lens' GetPersonTrackingResponse (Maybe VideoJobStatus)

The current status of the person tracking job.

getPersonTrackingResponse_persons :: Lens' GetPersonTrackingResponse (Maybe [PersonDetection])

An array of the persons detected in the video and the time(s) their path was tracked throughout the video. An array element will exist for each time a person's path is tracked.

getPersonTrackingResponse_httpStatus :: Lens' GetPersonTrackingResponse Int

The response's http status code.