Network.AWS.Rekognition.ListCollections

Description

Returns list of collection IDs in your account. If the result is truncated, the response also provides a NextToken that you can use in the subsequent request to fetch the next set of collection IDs.

For an example, see Listing Collections in the Amazon Rekognition Developer Guide.

This operation requires permissions to perform the rekognition:ListCollections action.

This operation returns paginated results.

Creating a Request

data ListCollections

See: newListCollections smart constructor.

Constructors

ListCollections'

Fields

nextToken :: Maybe Text

Pagination token from the previous response.

maxResults :: Maybe Natural

Maximum number of collection IDs to return.

newListCollections :: ListCollections

Create a value of ListCollections with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListCollections', listCollections_nextToken - Pagination token from the previous response.

$sel:maxResults:ListCollections', listCollections_maxResults - Maximum number of collection IDs to return.

Request Lenses

listCollections_nextToken :: Lens' ListCollections (Maybe Text)

Pagination token from the previous response.

listCollections_maxResults :: Lens' ListCollections (Maybe Natural)

Maximum number of collection IDs to return.

Destructuring the Response

data ListCollectionsResponse

See: newListCollectionsResponse smart constructor.

Constructors

ListCollectionsResponse'

Fields

faceModelVersions :: Maybe [Text]

Version numbers of the face detection models associated with the collections in the array CollectionIds. For example, the value of FaceModelVersions[2] is the version number for the face detection model used by the collection in CollectionId[2].

nextToken :: Maybe Text

If the result is truncated, the response provides a NextToken that you can use in the subsequent request to fetch the next set of collection IDs.

collectionIds :: Maybe [Text]

An array of collection IDs.

httpStatus :: Int

The response's http status code.

newListCollectionsResponse

Arguments

:: Int
$sel:httpStatus:ListCollectionsResponse'

-> ListCollectionsResponse

Create a value of ListCollectionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:faceModelVersions:ListCollectionsResponse', listCollectionsResponse_faceModelVersions - Version numbers of the face detection models associated with the collections in the array CollectionIds. For example, the value of FaceModelVersions[2] is the version number for the face detection model used by the collection in CollectionId[2].

$sel:nextToken:ListCollections', listCollectionsResponse_nextToken - If the result is truncated, the response provides a NextToken that you can use in the subsequent request to fetch the next set of collection IDs.

$sel:collectionIds:ListCollectionsResponse', listCollectionsResponse_collectionIds - An array of collection IDs.

$sel:httpStatus:ListCollectionsResponse', listCollectionsResponse_httpStatus - The response's http status code.

Response Lenses

listCollectionsResponse_faceModelVersions :: Lens' ListCollectionsResponse (Maybe [Text])

Version numbers of the face detection models associated with the collections in the array CollectionIds. For example, the value of FaceModelVersions[2] is the version number for the face detection model used by the collection in CollectionId[2].

listCollectionsResponse_nextToken :: Lens' ListCollectionsResponse (Maybe Text)

If the result is truncated, the response provides a NextToken that you can use in the subsequent request to fetch the next set of collection IDs.

listCollectionsResponse_collectionIds :: Lens' ListCollectionsResponse (Maybe [Text])

An array of collection IDs.

listCollectionsResponse_httpStatus :: Lens' ListCollectionsResponse Int

The response's http status code.