Network.AWS.Rekognition.ListStreamProcessors

Description

Gets a list of stream processors that you have created with CreateStreamProcessor.

This operation returns paginated results.

Creating a Request

data ListStreamProcessors

See: newListStreamProcessors smart constructor.

Constructors

ListStreamProcessors'

Fields

nextToken :: Maybe Text

If the previous response was incomplete (because there are more stream processors to retrieve), Amazon Rekognition Video returns a pagination token in the response. You can use this pagination token to retrieve the next set of stream processors.

maxResults :: Maybe Natural

Maximum number of stream processors you want Amazon Rekognition Video to return in the response. The default is 1000.

newListStreamProcessors :: ListStreamProcessors

Create a value of ListStreamProcessors with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListStreamProcessors', listStreamProcessors_nextToken - If the previous response was incomplete (because there are more stream processors to retrieve), Amazon Rekognition Video returns a pagination token in the response. You can use this pagination token to retrieve the next set of stream processors.

$sel:maxResults:ListStreamProcessors', listStreamProcessors_maxResults - Maximum number of stream processors you want Amazon Rekognition Video to return in the response. The default is 1000.

Request Lenses

listStreamProcessors_nextToken :: Lens' ListStreamProcessors (Maybe Text)

If the previous response was incomplete (because there are more stream processors to retrieve), Amazon Rekognition Video returns a pagination token in the response. You can use this pagination token to retrieve the next set of stream processors.

listStreamProcessors_maxResults :: Lens' ListStreamProcessors (Maybe Natural)

Maximum number of stream processors you want Amazon Rekognition Video to return in the response. The default is 1000.

Destructuring the Response

data ListStreamProcessorsResponse

See: newListStreamProcessorsResponse smart constructor.

Constructors

ListStreamProcessorsResponse'

Fields

streamProcessors :: Maybe [StreamProcessor]

List of stream processors that you have created.

nextToken :: Maybe Text

If the response is truncated, Amazon Rekognition Video returns this token that you can use in the subsequent request to retrieve the next set of stream processors.

httpStatus :: Int

The response's http status code.

newListStreamProcessorsResponse

Arguments

:: Int
$sel:httpStatus:ListStreamProcessorsResponse'

-> ListStreamProcessorsResponse

Create a value of ListStreamProcessorsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamProcessors:ListStreamProcessorsResponse', listStreamProcessorsResponse_streamProcessors - List of stream processors that you have created.

$sel:nextToken:ListStreamProcessors', listStreamProcessorsResponse_nextToken - If the response is truncated, Amazon Rekognition Video returns this token that you can use in the subsequent request to retrieve the next set of stream processors.

$sel:httpStatus:ListStreamProcessorsResponse', listStreamProcessorsResponse_httpStatus - The response's http status code.

Response Lenses

listStreamProcessorsResponse_streamProcessors :: Lens' ListStreamProcessorsResponse (Maybe [StreamProcessor])

List of stream processors that you have created.

listStreamProcessorsResponse_nextToken :: Lens' ListStreamProcessorsResponse (Maybe Text)

If the response is truncated, Amazon Rekognition Video returns this token that you can use in the subsequent request to retrieve the next set of stream processors.

listStreamProcessorsResponse_httpStatus :: Lens' ListStreamProcessorsResponse Int

The response's http status code.