Network.AWS.Rekognition.SearchFaces

Description

For a given input face ID, searches for matching faces in the collection the face belongs to. You get a face ID when you add a face to the collection using the IndexFaces operation. The operation compares the features of the input face with faces in the specified collection.

You can also search faces without indexing faces by using the SearchFacesByImage operation.

The operation response returns an array of faces that match, ordered by similarity score with the highest similarity first. More specifically, it is an array of metadata for each face match that is found. Along with the metadata, the response also includes a confidence value for each face match, indicating the confidence that the specific face matches the input face.

For an example, see Searching for a Face Using Its Face ID in the Amazon Rekognition Developer Guide.

This operation requires permissions to perform the rekognition:SearchFaces action.

Creating a Request

data SearchFaces

See: newSearchFaces smart constructor.

Constructors

SearchFaces'

Fields

maxFaces :: Maybe Natural

Maximum number of faces to return. The operation returns the maximum number of faces with the highest confidence in the match.

faceMatchThreshold :: Maybe Double

Optional value specifying the minimum confidence in the face match to return. For example, don't return any matches where confidence in matches is less than 70%. The default value is 80%.

collectionId :: Text

ID of the collection the face belongs to.

faceId :: Text

ID of a face to find matches for in the collection.

newSearchFaces

Arguments

:: Text
$sel:collectionId:SearchFaces'

-> Text
$sel:faceId:SearchFaces'

-> SearchFaces

Create a value of SearchFaces with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxFaces:SearchFaces', searchFaces_maxFaces - Maximum number of faces to return. The operation returns the maximum number of faces with the highest confidence in the match.

$sel:faceMatchThreshold:SearchFaces', searchFaces_faceMatchThreshold - Optional value specifying the minimum confidence in the face match to return. For example, don't return any matches where confidence in matches is less than 70%. The default value is 80%.

$sel:collectionId:SearchFaces', searchFaces_collectionId - ID of the collection the face belongs to.

$sel:faceId:SearchFaces', searchFaces_faceId - ID of a face to find matches for in the collection.

Request Lenses

searchFaces_maxFaces :: Lens' SearchFaces (Maybe Natural)

Maximum number of faces to return. The operation returns the maximum number of faces with the highest confidence in the match.

searchFaces_faceMatchThreshold :: Lens' SearchFaces (Maybe Double)

Optional value specifying the minimum confidence in the face match to return. For example, don't return any matches where confidence in matches is less than 70%. The default value is 80%.

searchFaces_collectionId :: Lens' SearchFaces Text

ID of the collection the face belongs to.

searchFaces_faceId :: Lens' SearchFaces Text

ID of a face to find matches for in the collection.

Destructuring the Response

data SearchFacesResponse

See: newSearchFacesResponse smart constructor.

Constructors

SearchFacesResponse'

Fields

faceModelVersion :: Maybe Text

Version number of the face detection model associated with the input collection (CollectionId).

faceMatches :: Maybe [FaceMatch]

An array of faces that matched the input face, along with the confidence in the match.

searchedFaceId :: Maybe Text

ID of the face that was searched for matches in a collection.

httpStatus :: Int

The response's http status code.

newSearchFacesResponse

Arguments

:: Int
$sel:httpStatus:SearchFacesResponse'

-> SearchFacesResponse

Create a value of SearchFacesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:faceModelVersion:SearchFacesResponse', searchFacesResponse_faceModelVersion - Version number of the face detection model associated with the input collection (CollectionId).

$sel:faceMatches:SearchFacesResponse', searchFacesResponse_faceMatches - An array of faces that matched the input face, along with the confidence in the match.

$sel:searchedFaceId:SearchFacesResponse', searchFacesResponse_searchedFaceId - ID of the face that was searched for matches in a collection.

$sel:httpStatus:SearchFacesResponse', searchFacesResponse_httpStatus - The response's http status code.

Response Lenses

searchFacesResponse_faceModelVersion :: Lens' SearchFacesResponse (Maybe Text)

Version number of the face detection model associated with the input collection (CollectionId).

searchFacesResponse_faceMatches :: Lens' SearchFacesResponse (Maybe [FaceMatch])

An array of faces that matched the input face, along with the confidence in the match.

searchFacesResponse_searchedFaceId :: Lens' SearchFacesResponse (Maybe Text)

ID of the face that was searched for matches in a collection.

searchFacesResponse_httpStatus :: Lens' SearchFacesResponse Int

The response's http status code.