Network.AWS.Rekognition.StartCelebrityRecognition

Description

Starts asynchronous recognition of celebrities in a stored video.

Amazon Rekognition Video can detect celebrities in a video must be stored in an Amazon S3 bucket. Use Video to specify the bucket name and the filename of the video. StartCelebrityRecognition returns a job identifier (JobId) which you use to get the results of the analysis. When celebrity recognition analysis is finished, Amazon Rekognition Video publishes a completion status to the Amazon Simple Notification Service topic that you specify in NotificationChannel. To get the results of the celebrity recognition analysis, first check that the status value published to the Amazon SNS topic is SUCCEEDED. If so, call GetCelebrityRecognition and pass the job identifier (JobId) from the initial call to StartCelebrityRecognition.

For more information, see Recognizing Celebrities in the Amazon Rekognition Developer Guide.

Creating a Request

data StartCelebrityRecognition

See: newStartCelebrityRecognition smart constructor.

Constructors

StartCelebrityRecognition'

Fields

notificationChannel :: Maybe NotificationChannel

The Amazon SNS topic ARN that you want Amazon Rekognition Video to publish the completion status of the celebrity recognition analysis to. The Amazon SNS topic must have a topic name that begins with AmazonRekognition if you are using the AmazonRekognitionServiceRole permissions policy.

clientRequestToken :: Maybe Text

Idempotent token used to identify the start request. If you use the same token with multiple StartCelebrityRecognition requests, the same JobId is returned. Use ClientRequestToken to prevent the same job from being accidently started more than once.

jobTag :: Maybe Text

An identifier you specify that's returned in the completion notification that's published to your Amazon Simple Notification Service topic. For example, you can use JobTag to group related jobs and identify them in the completion notification.

video :: Video

The video in which you want to recognize celebrities. The video must be stored in an Amazon S3 bucket.

newStartCelebrityRecognition

Arguments

:: Video
$sel:video:StartCelebrityRecognition'

-> StartCelebrityRecognition

Create a value of StartCelebrityRecognition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notificationChannel:StartCelebrityRecognition', startCelebrityRecognition_notificationChannel - The Amazon SNS topic ARN that you want Amazon Rekognition Video to publish the completion status of the celebrity recognition analysis to. The Amazon SNS topic must have a topic name that begins with AmazonRekognition if you are using the AmazonRekognitionServiceRole permissions policy.

$sel:clientRequestToken:StartCelebrityRecognition', startCelebrityRecognition_clientRequestToken - Idempotent token used to identify the start request. If you use the same token with multiple StartCelebrityRecognition requests, the same JobId is returned. Use ClientRequestToken to prevent the same job from being accidently started more than once.

$sel:jobTag:StartCelebrityRecognition', startCelebrityRecognition_jobTag - An identifier you specify that's returned in the completion notification that's published to your Amazon Simple Notification Service topic. For example, you can use JobTag to group related jobs and identify them in the completion notification.

$sel:video:StartCelebrityRecognition', startCelebrityRecognition_video - The video in which you want to recognize celebrities. The video must be stored in an Amazon S3 bucket.

Request Lenses

startCelebrityRecognition_notificationChannel :: Lens' StartCelebrityRecognition (Maybe NotificationChannel)

The Amazon SNS topic ARN that you want Amazon Rekognition Video to publish the completion status of the celebrity recognition analysis to. The Amazon SNS topic must have a topic name that begins with AmazonRekognition if you are using the AmazonRekognitionServiceRole permissions policy.

startCelebrityRecognition_clientRequestToken :: Lens' StartCelebrityRecognition (Maybe Text)

Idempotent token used to identify the start request. If you use the same token with multiple StartCelebrityRecognition requests, the same JobId is returned. Use ClientRequestToken to prevent the same job from being accidently started more than once.

startCelebrityRecognition_jobTag :: Lens' StartCelebrityRecognition (Maybe Text)

An identifier you specify that's returned in the completion notification that's published to your Amazon Simple Notification Service topic. For example, you can use JobTag to group related jobs and identify them in the completion notification.

startCelebrityRecognition_video :: Lens' StartCelebrityRecognition Video

The video in which you want to recognize celebrities. The video must be stored in an Amazon S3 bucket.

Destructuring the Response

data StartCelebrityRecognitionResponse

See: newStartCelebrityRecognitionResponse smart constructor.

Constructors

StartCelebrityRecognitionResponse'

Fields

jobId :: Maybe Text

The identifier for the celebrity recognition analysis job. Use JobId to identify the job in a subsequent call to GetCelebrityRecognition.

httpStatus :: Int

The response's http status code.

newStartCelebrityRecognitionResponse

Arguments

:: Int
$sel:httpStatus:StartCelebrityRecognitionResponse'

-> StartCelebrityRecognitionResponse

Create a value of StartCelebrityRecognitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:StartCelebrityRecognitionResponse', startCelebrityRecognitionResponse_jobId - The identifier for the celebrity recognition analysis job. Use JobId to identify the job in a subsequent call to GetCelebrityRecognition.

$sel:httpStatus:StartCelebrityRecognitionResponse', startCelebrityRecognitionResponse_httpStatus - The response's http status code.

Response Lenses

startCelebrityRecognitionResponse_jobId :: Lens' StartCelebrityRecognitionResponse (Maybe Text)

The identifier for the celebrity recognition analysis job. Use JobId to identify the job in a subsequent call to GetCelebrityRecognition.

startCelebrityRecognitionResponse_httpStatus :: Lens' StartCelebrityRecognitionResponse Int

The response's http status code.