Network.AWS.Rekognition.StartContentModeration

Description

Starts asynchronous detection of inappropriate, unwanted, or offensive content in a stored video. For a list of moderation labels in Amazon Rekognition, see Using the image and video moderation APIs.

Amazon Rekognition Video can moderate content in a video stored in an Amazon S3 bucket. Use Video to specify the bucket name and the filename of the video. StartContentModeration returns a job identifier (JobId) which you use to get the results of the analysis. When content analysis is finished, Amazon Rekognition Video publishes a completion status to the Amazon Simple Notification Service topic that you specify in NotificationChannel.

To get the results of the content analysis, first check that the status value published to the Amazon SNS topic is SUCCEEDED. If so, call GetContentModeration and pass the job identifier (JobId) from the initial call to StartContentModeration.

For more information, see Content moderation in the Amazon Rekognition Developer Guide.

Creating a Request

data StartContentModeration

See: newStartContentModeration smart constructor.

Constructors

StartContentModeration'

Fields

jobTag :: Maybe Text

An identifier you specify that's returned in the completion notification that's published to your Amazon Simple Notification Service topic. For example, you can use JobTag to group related jobs and identify them in the completion notification.

notificationChannel :: Maybe NotificationChannel

The Amazon SNS topic ARN that you want Amazon Rekognition Video to publish the completion status of the content analysis to. The Amazon SNS topic must have a topic name that begins with AmazonRekognition if you are using the AmazonRekognitionServiceRole permissions policy to access the topic.

clientRequestToken :: Maybe Text

Idempotent token used to identify the start request. If you use the same token with multiple StartContentModeration requests, the same JobId is returned. Use ClientRequestToken to prevent the same job from being accidently started more than once.

minConfidence :: Maybe Double

Specifies the minimum confidence that Amazon Rekognition must have in order to return a moderated content label. Confidence represents how certain Amazon Rekognition is that the moderated content is correctly identified. 0 is the lowest confidence. 100 is the highest confidence. Amazon Rekognition doesn't return any moderated content labels with a confidence level lower than this specified value. If you don't specify MinConfidence, GetContentModeration returns labels with confidence values greater than or equal to 50 percent.

video :: Video

The video in which you want to detect inappropriate, unwanted, or offensive content. The video must be stored in an Amazon S3 bucket.

newStartContentModeration

Arguments

:: Video
$sel:video:StartContentModeration'

-> StartContentModeration

Create a value of StartContentModeration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobTag:StartContentModeration', startContentModeration_jobTag - An identifier you specify that's returned in the completion notification that's published to your Amazon Simple Notification Service topic. For example, you can use JobTag to group related jobs and identify them in the completion notification.

$sel:notificationChannel:StartContentModeration', startContentModeration_notificationChannel - The Amazon SNS topic ARN that you want Amazon Rekognition Video to publish the completion status of the content analysis to. The Amazon SNS topic must have a topic name that begins with AmazonRekognition if you are using the AmazonRekognitionServiceRole permissions policy to access the topic.

$sel:clientRequestToken:StartContentModeration', startContentModeration_clientRequestToken - Idempotent token used to identify the start request. If you use the same token with multiple StartContentModeration requests, the same JobId is returned. Use ClientRequestToken to prevent the same job from being accidently started more than once.

$sel:minConfidence:StartContentModeration', startContentModeration_minConfidence - Specifies the minimum confidence that Amazon Rekognition must have in order to return a moderated content label. Confidence represents how certain Amazon Rekognition is that the moderated content is correctly identified. 0 is the lowest confidence. 100 is the highest confidence. Amazon Rekognition doesn't return any moderated content labels with a confidence level lower than this specified value. If you don't specify MinConfidence, GetContentModeration returns labels with confidence values greater than or equal to 50 percent.

$sel:video:StartContentModeration', startContentModeration_video - The video in which you want to detect inappropriate, unwanted, or offensive content. The video must be stored in an Amazon S3 bucket.

Request Lenses

startContentModeration_jobTag :: Lens' StartContentModeration (Maybe Text)

An identifier you specify that's returned in the completion notification that's published to your Amazon Simple Notification Service topic. For example, you can use JobTag to group related jobs and identify them in the completion notification.

startContentModeration_notificationChannel :: Lens' StartContentModeration (Maybe NotificationChannel)

The Amazon SNS topic ARN that you want Amazon Rekognition Video to publish the completion status of the content analysis to. The Amazon SNS topic must have a topic name that begins with AmazonRekognition if you are using the AmazonRekognitionServiceRole permissions policy to access the topic.

startContentModeration_clientRequestToken :: Lens' StartContentModeration (Maybe Text)

Idempotent token used to identify the start request. If you use the same token with multiple StartContentModeration requests, the same JobId is returned. Use ClientRequestToken to prevent the same job from being accidently started more than once.

startContentModeration_minConfidence :: Lens' StartContentModeration (Maybe Double)

Specifies the minimum confidence that Amazon Rekognition must have in order to return a moderated content label. Confidence represents how certain Amazon Rekognition is that the moderated content is correctly identified. 0 is the lowest confidence. 100 is the highest confidence. Amazon Rekognition doesn't return any moderated content labels with a confidence level lower than this specified value. If you don't specify MinConfidence, GetContentModeration returns labels with confidence values greater than or equal to 50 percent.

startContentModeration_video :: Lens' StartContentModeration Video

The video in which you want to detect inappropriate, unwanted, or offensive content. The video must be stored in an Amazon S3 bucket.

Destructuring the Response

data StartContentModerationResponse

See: newStartContentModerationResponse smart constructor.

Constructors

StartContentModerationResponse'

Fields

jobId :: Maybe Text

The identifier for the content analysis job. Use JobId to identify the job in a subsequent call to GetContentModeration.

httpStatus :: Int

The response's http status code.

newStartContentModerationResponse

Arguments

:: Int
$sel:httpStatus:StartContentModerationResponse'

-> StartContentModerationResponse

Create a value of StartContentModerationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:StartContentModerationResponse', startContentModerationResponse_jobId - The identifier for the content analysis job. Use JobId to identify the job in a subsequent call to GetContentModeration.

$sel:httpStatus:StartContentModerationResponse', startContentModerationResponse_httpStatus - The response's http status code.

Response Lenses

startContentModerationResponse_jobId :: Lens' StartContentModerationResponse (Maybe Text)

The identifier for the content analysis job. Use JobId to identify the job in a subsequent call to GetContentModeration.

startContentModerationResponse_httpStatus :: Lens' StartContentModerationResponse Int

The response's http status code.