Network.AWS.Rekognition.StartProjectVersion

Description

Starts the running of the version of a model. Starting a model takes a while to complete. To check the current state of the model, use DescribeProjectVersions.

Once the model is running, you can detect custom labels in new images by calling DetectCustomLabels.

You are charged for the amount of time that the model is running. To stop a running model, call StopProjectVersion.

This operation requires permissions to perform the rekognition:StartProjectVersion action.

Creating a Request

data StartProjectVersion

See: newStartProjectVersion smart constructor.

Constructors

StartProjectVersion'

Fields

projectVersionArn :: Text

The Amazon Resource Name(ARN) of the model version that you want to start.

minInferenceUnits :: Natural

The minimum number of inference units to use. A single inference unit represents 1 hour of processing and can support up to 5 Transaction Pers Second (TPS). Use a higher number to increase the TPS throughput of your model. You are charged for the number of inference units that you use.

newStartProjectVersion

Arguments

:: Text
$sel:projectVersionArn:StartProjectVersion'

-> Natural
$sel:minInferenceUnits:StartProjectVersion'

-> StartProjectVersion

Create a value of StartProjectVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectVersionArn:StartProjectVersion', startProjectVersion_projectVersionArn - The Amazon Resource Name(ARN) of the model version that you want to start.

$sel:minInferenceUnits:StartProjectVersion', startProjectVersion_minInferenceUnits - The minimum number of inference units to use. A single inference unit represents 1 hour of processing and can support up to 5 Transaction Pers Second (TPS). Use a higher number to increase the TPS throughput of your model. You are charged for the number of inference units that you use.

Request Lenses

startProjectVersion_projectVersionArn :: Lens' StartProjectVersion Text

The Amazon Resource Name(ARN) of the model version that you want to start.

startProjectVersion_minInferenceUnits :: Lens' StartProjectVersion Natural

The minimum number of inference units to use. A single inference unit represents 1 hour of processing and can support up to 5 Transaction Pers Second (TPS). Use a higher number to increase the TPS throughput of your model. You are charged for the number of inference units that you use.

Destructuring the Response

data StartProjectVersionResponse

See: newStartProjectVersionResponse smart constructor.

Constructors

StartProjectVersionResponse'

Fields

status :: Maybe ProjectVersionStatus

The current running status of the model.

httpStatus :: Int

The response's http status code.

newStartProjectVersionResponse

Arguments

:: Int
$sel:httpStatus:StartProjectVersionResponse'

-> StartProjectVersionResponse

Create a value of StartProjectVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:StartProjectVersionResponse', startProjectVersionResponse_status - The current running status of the model.

$sel:httpStatus:StartProjectVersionResponse', startProjectVersionResponse_httpStatus - The response's http status code.

Response Lenses

startProjectVersionResponse_status :: Lens' StartProjectVersionResponse (Maybe ProjectVersionStatus)

The current running status of the model.

startProjectVersionResponse_httpStatus :: Lens' StartProjectVersionResponse Int

The response's http status code.