Network.AWS.Rekognition.StartSegmentDetection

Description

Starts asynchronous detection of segment detection in a stored video.

Amazon Rekognition Video can detect segments in a video stored in an Amazon S3 bucket. Use Video to specify the bucket name and the filename of the video. StartSegmentDetection returns a job identifier (JobId) which you use to get the results of the operation. When segment detection is finished, Amazon Rekognition Video publishes a completion status to the Amazon Simple Notification Service topic that you specify in NotificationChannel.

You can use the Filters (StartSegmentDetectionFilters) input parameter to specify the minimum detection confidence returned in the response. Within Filters, use ShotFilter (StartShotDetectionFilter) to filter detected shots. Use TechnicalCueFilter (StartTechnicalCueDetectionFilter) to filter technical cues.

To get the results of the segment detection operation, first check that the status value published to the Amazon SNS topic is SUCCEEDED. if so, call GetSegmentDetection and pass the job identifier (JobId) from the initial call to StartSegmentDetection.

For more information, see Detecting Video Segments in Stored Video in the Amazon Rekognition Developer Guide.

Creating a Request

data StartSegmentDetection

See: newStartSegmentDetection smart constructor.

Constructors

StartSegmentDetection'

Fields

jobTag :: Maybe Text

An identifier you specify that's returned in the completion notification that's published to your Amazon Simple Notification Service topic. For example, you can use JobTag to group related jobs and identify them in the completion notification.

filters :: Maybe StartSegmentDetectionFilters

Filters for technical cue or shot detection.

notificationChannel :: Maybe NotificationChannel

The ARN of the Amazon SNS topic to which you want Amazon Rekognition Video to publish the completion status of the segment detection operation. Note that the Amazon SNS topic must have a topic name that begins with AmazonRekognition if you are using the AmazonRekognitionServiceRole permissions policy to access the topic.

clientRequestToken :: Maybe Text

Idempotent token used to identify the start request. If you use the same token with multiple StartSegmentDetection requests, the same JobId is returned. Use ClientRequestToken to prevent the same job from being accidently started more than once.

video :: Video

segmentTypes :: NonEmpty SegmentType

An array of segment types to detect in the video. Valid values are TECHNICAL_CUE and SHOT.

newStartSegmentDetection

Arguments

:: Video
$sel:video:StartSegmentDetection'

-> NonEmpty SegmentType
$sel:segmentTypes:StartSegmentDetection'

-> StartSegmentDetection

Create a value of StartSegmentDetection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobTag:StartSegmentDetection', startSegmentDetection_jobTag - An identifier you specify that's returned in the completion notification that's published to your Amazon Simple Notification Service topic. For example, you can use JobTag to group related jobs and identify them in the completion notification.

$sel:filters:StartSegmentDetection', startSegmentDetection_filters - Filters for technical cue or shot detection.

$sel:notificationChannel:StartSegmentDetection', startSegmentDetection_notificationChannel - The ARN of the Amazon SNS topic to which you want Amazon Rekognition Video to publish the completion status of the segment detection operation. Note that the Amazon SNS topic must have a topic name that begins with AmazonRekognition if you are using the AmazonRekognitionServiceRole permissions policy to access the topic.

$sel:clientRequestToken:StartSegmentDetection', startSegmentDetection_clientRequestToken - Idempotent token used to identify the start request. If you use the same token with multiple StartSegmentDetection requests, the same JobId is returned. Use ClientRequestToken to prevent the same job from being accidently started more than once.

$sel:video:StartSegmentDetection', startSegmentDetection_video - Undocumented member.

$sel:segmentTypes:StartSegmentDetection', startSegmentDetection_segmentTypes - An array of segment types to detect in the video. Valid values are TECHNICAL_CUE and SHOT.

Request Lenses

startSegmentDetection_jobTag :: Lens' StartSegmentDetection (Maybe Text)

An identifier you specify that's returned in the completion notification that's published to your Amazon Simple Notification Service topic. For example, you can use JobTag to group related jobs and identify them in the completion notification.

startSegmentDetection_filters :: Lens' StartSegmentDetection (Maybe StartSegmentDetectionFilters)

Filters for technical cue or shot detection.

startSegmentDetection_notificationChannel :: Lens' StartSegmentDetection (Maybe NotificationChannel)

The ARN of the Amazon SNS topic to which you want Amazon Rekognition Video to publish the completion status of the segment detection operation. Note that the Amazon SNS topic must have a topic name that begins with AmazonRekognition if you are using the AmazonRekognitionServiceRole permissions policy to access the topic.

startSegmentDetection_clientRequestToken :: Lens' StartSegmentDetection (Maybe Text)

Idempotent token used to identify the start request. If you use the same token with multiple StartSegmentDetection requests, the same JobId is returned. Use ClientRequestToken to prevent the same job from being accidently started more than once.

startSegmentDetection_video :: Lens' StartSegmentDetection Video

Undocumented member.

startSegmentDetection_segmentTypes :: Lens' StartSegmentDetection (NonEmpty SegmentType)

An array of segment types to detect in the video. Valid values are TECHNICAL_CUE and SHOT.

Destructuring the Response

data StartSegmentDetectionResponse

See: newStartSegmentDetectionResponse smart constructor.

Constructors

StartSegmentDetectionResponse'

Fields

jobId :: Maybe Text

Unique identifier for the segment detection job. The JobId is returned from StartSegmentDetection.

httpStatus :: Int

The response's http status code.

newStartSegmentDetectionResponse

Arguments

:: Int
$sel:httpStatus:StartSegmentDetectionResponse'

-> StartSegmentDetectionResponse

Create a value of StartSegmentDetectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:StartSegmentDetectionResponse', startSegmentDetectionResponse_jobId - Unique identifier for the segment detection job. The JobId is returned from StartSegmentDetection.

$sel:httpStatus:StartSegmentDetectionResponse', startSegmentDetectionResponse_httpStatus - The response's http status code.

Response Lenses

startSegmentDetectionResponse_jobId :: Lens' StartSegmentDetectionResponse (Maybe Text)

Unique identifier for the segment detection job. The JobId is returned from StartSegmentDetection.

startSegmentDetectionResponse_httpStatus :: Lens' StartSegmentDetectionResponse Int

The response's http status code.