Network.AWS.Rekognition.Types.Asset

Description

Documentation

data Asset

Assets are the images that you use to train and evaluate a model version. Assets can also contain validation information that you use to debug a failed model training.

See: newAsset smart constructor.

Constructors

Asset'

Fields

groundTruthManifest :: Maybe GroundTruthManifest

newAsset :: Asset

Create a value of Asset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groundTruthManifest:Asset', asset_groundTruthManifest - Undocumented member.

asset_groundTruthManifest :: Lens' Asset (Maybe GroundTruthManifest)

Undocumented member.