Network.AWS.Rekognition.Types.Beard

Description

Documentation

data Beard

Indicates whether or not the face has a beard, and the confidence level in the determination.

See: newBeard smart constructor.

Constructors

Beard'

Fields

confidence :: Maybe Double

Level of confidence in the determination.

value :: Maybe Bool

Boolean value that indicates whether the face has beard or not.

newBeard :: Beard

Create a value of Beard with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:confidence:Beard', beard_confidence - Level of confidence in the determination.

$sel:value:Beard', beard_value - Boolean value that indicates whether the face has beard or not.

beard_confidence :: Lens' Beard (Maybe Double)

Level of confidence in the determination.

beard_value :: Lens' Beard (Maybe Bool)

Boolean value that indicates whether the face has beard or not.