Network.AWS.Rekognition.Types.BoundingBox

Description

Documentation

data BoundingBox

Identifies the bounding box around the label, face, text or personal protective equipment. The left (x-coordinate) and top (y-coordinate) are coordinates representing the top and left sides of the bounding box. Note that the upper-left corner of the image is the origin (0,0).

The top and left values returned are ratios of the overall image size. For example, if the input image is 700x200 pixels, and the top-left coordinate of the bounding box is 350x50 pixels, the API returns a left value of 0.5 (350/700) and a top value of 0.25 (50/200).

The width and height values represent the dimensions of the bounding box as a ratio of the overall image dimension. For example, if the input image is 700x200 pixels, and the bounding box width is 70 pixels, the width returned is 0.1.

The bounding box coordinates can have negative values. For example, if Amazon Rekognition is able to detect a face that is at the image edge and is only partially visible, the service can return coordinates that are outside the image bounds and, depending on the image edge, you might get negative values or values greater than 1 for the left or top values.

See: newBoundingBox smart constructor.

Constructors

BoundingBox'

Fields

height :: Maybe Double

Height of the bounding box as a ratio of the overall image height.

width :: Maybe Double

Width of the bounding box as a ratio of the overall image width.

left :: Maybe Double

Left coordinate of the bounding box as a ratio of overall image width.

top :: Maybe Double

Top coordinate of the bounding box as a ratio of overall image height.

newBoundingBox :: BoundingBox

Create a value of BoundingBox with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:height:BoundingBox', boundingBox_height - Height of the bounding box as a ratio of the overall image height.

$sel:width:BoundingBox', boundingBox_width - Width of the bounding box as a ratio of the overall image width.

$sel:left:BoundingBox', boundingBox_left - Left coordinate of the bounding box as a ratio of overall image width.

$sel:top:BoundingBox', boundingBox_top - Top coordinate of the bounding box as a ratio of overall image height.

boundingBox_height :: Lens' BoundingBox (Maybe Double)

Height of the bounding box as a ratio of the overall image height.

boundingBox_width :: Lens' BoundingBox (Maybe Double)

Width of the bounding box as a ratio of the overall image width.

boundingBox_left :: Lens' BoundingBox (Maybe Double)

Left coordinate of the bounding box as a ratio of overall image width.

boundingBox_top :: Lens' BoundingBox (Maybe Double)

Top coordinate of the bounding box as a ratio of overall image height.