Network.AWS.Rekognition.Types.CelebrityRecognition

Description

Documentation

data CelebrityRecognition

Information about a detected celebrity and the time the celebrity was detected in a stored video. For more information, see GetCelebrityRecognition in the Amazon Rekognition Developer Guide.

See: newCelebrityRecognition smart constructor.

Constructors

CelebrityRecognition'

Fields

timestamp :: Maybe Integer

The time, in milliseconds from the start of the video, that the celebrity was recognized.

celebrity :: Maybe CelebrityDetail

Information about a recognized celebrity.

newCelebrityRecognition :: CelebrityRecognition

Create a value of CelebrityRecognition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timestamp:CelebrityRecognition', celebrityRecognition_timestamp - The time, in milliseconds from the start of the video, that the celebrity was recognized.

$sel:celebrity:CelebrityRecognition', celebrityRecognition_celebrity - Information about a recognized celebrity.

celebrityRecognition_timestamp :: Lens' CelebrityRecognition (Maybe Integer)

The time, in milliseconds from the start of the video, that the celebrity was recognized.

celebrityRecognition_celebrity :: Lens' CelebrityRecognition (Maybe CelebrityDetail)

Information about a recognized celebrity.