Network.AWS.Rekognition.Types.CompareFacesMatch

Description

Documentation

data CompareFacesMatch

Provides information about a face in a target image that matches the source image face analyzed by CompareFaces. The Face property contains the bounding box of the face in the target image. The Similarity property is the confidence that the source image face matches the face in the bounding box.

See: newCompareFacesMatch smart constructor.

Constructors

CompareFacesMatch'

Fields

similarity :: Maybe Double

Level of confidence that the faces match.

face :: Maybe ComparedFace

Provides face metadata (bounding box and confidence that the bounding box actually contains a face).

newCompareFacesMatch :: CompareFacesMatch

Create a value of CompareFacesMatch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:similarity:CompareFacesMatch', compareFacesMatch_similarity - Level of confidence that the faces match.

$sel:face:CompareFacesMatch', compareFacesMatch_face - Provides face metadata (bounding box and confidence that the bounding box actually contains a face).

compareFacesMatch_similarity :: Lens' CompareFacesMatch (Maybe Double)

Level of confidence that the faces match.

compareFacesMatch_face :: Lens' CompareFacesMatch (Maybe ComparedFace)

Provides face metadata (bounding box and confidence that the bounding box actually contains a face).