Network.AWS.Rekognition.Types.ComparedFace

Description

Documentation

data ComparedFace

Provides face metadata for target image faces that are analyzed by CompareFaces and RecognizeCelebrities.

See: newComparedFace smart constructor.

Constructors

ComparedFace'

Fields

landmarks :: Maybe [Landmark]

An array of facial landmarks.

pose :: Maybe Pose

Indicates the pose of the face as determined by its pitch, roll, and yaw.

emotions :: Maybe [Emotion]

The emotions that appear to be expressed on the face, and the confidence level in the determination. Valid values include "Happy", "Sad", "Angry", "Confused", "Disgusted", "Surprised", "Calm", "Unknown", and "Fear".

boundingBox :: Maybe BoundingBox

Bounding box of the face.

confidence :: Maybe Double

Level of confidence that what the bounding box contains is a face.

smile :: Maybe Smile

Indicates whether or not the face is smiling, and the confidence level in the determination.

quality :: Maybe ImageQuality

Identifies face image brightness and sharpness.

newComparedFace :: ComparedFace

Create a value of ComparedFace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:landmarks:ComparedFace', comparedFace_landmarks - An array of facial landmarks.

$sel:pose:ComparedFace', comparedFace_pose - Indicates the pose of the face as determined by its pitch, roll, and yaw.

$sel:emotions:ComparedFace', comparedFace_emotions - The emotions that appear to be expressed on the face, and the confidence level in the determination. Valid values include "Happy", "Sad", "Angry", "Confused", "Disgusted", "Surprised", "Calm", "Unknown", and "Fear".

$sel:boundingBox:ComparedFace', comparedFace_boundingBox - Bounding box of the face.

$sel:confidence:ComparedFace', comparedFace_confidence - Level of confidence that what the bounding box contains is a face.

$sel:smile:ComparedFace', comparedFace_smile - Indicates whether or not the face is smiling, and the confidence level in the determination.

$sel:quality:ComparedFace', comparedFace_quality - Identifies face image brightness and sharpness.

comparedFace_landmarks :: Lens' ComparedFace (Maybe [Landmark])

An array of facial landmarks.

comparedFace_pose :: Lens' ComparedFace (Maybe Pose)

Indicates the pose of the face as determined by its pitch, roll, and yaw.

comparedFace_emotions :: Lens' ComparedFace (Maybe [Emotion])

The emotions that appear to be expressed on the face, and the confidence level in the determination. Valid values include "Happy", "Sad", "Angry", "Confused", "Disgusted", "Surprised", "Calm", "Unknown", and "Fear".

comparedFace_boundingBox :: Lens' ComparedFace (Maybe BoundingBox)

Bounding box of the face.

comparedFace_confidence :: Lens' ComparedFace (Maybe Double)

Level of confidence that what the bounding box contains is a face.

comparedFace_smile :: Lens' ComparedFace (Maybe Smile)

Indicates whether or not the face is smiling, and the confidence level in the determination.

comparedFace_quality :: Lens' ComparedFace (Maybe ImageQuality)

Identifies face image brightness and sharpness.