Network.AWS.Rekognition.Types.Face

Description

Documentation

data Face

Describes the face properties such as the bounding box, face ID, image ID of the input image, and external image ID that you assigned.

See: newFace smart constructor.

Constructors

Face'

Fields

faceId :: Maybe Text

Unique identifier that Amazon Rekognition assigns to the face.

imageId :: Maybe Text

Unique identifier that Amazon Rekognition assigns to the input image.

externalImageId :: Maybe Text

Identifier that you assign to all the faces in the input image.

boundingBox :: Maybe BoundingBox

Bounding box of the face.

confidence :: Maybe Double

Confidence level that the bounding box contains a face (and not a different object such as a tree).

newFace :: Face

Create a value of Face with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:faceId:Face', face_faceId - Unique identifier that Amazon Rekognition assigns to the face.

$sel:imageId:Face', face_imageId - Unique identifier that Amazon Rekognition assigns to the input image.

$sel:externalImageId:Face', face_externalImageId - Identifier that you assign to all the faces in the input image.

$sel:boundingBox:Face', face_boundingBox - Bounding box of the face.

$sel:confidence:Face', face_confidence - Confidence level that the bounding box contains a face (and not a different object such as a tree).

face_faceId :: Lens' Face (Maybe Text)

Unique identifier that Amazon Rekognition assigns to the face.

face_imageId :: Lens' Face (Maybe Text)

Unique identifier that Amazon Rekognition assigns to the input image.

face_externalImageId :: Lens' Face (Maybe Text)

Identifier that you assign to all the faces in the input image.

face_boundingBox :: Lens' Face (Maybe BoundingBox)

Bounding box of the face.

face_confidence :: Lens' Face (Maybe Double)

Confidence level that the bounding box contains a face (and not a different object such as a tree).