Network.AWS.Rekognition.Types.FaceMatch

Description

Documentation

data FaceMatch

Provides face metadata. In addition, it also provides the confidence in the match of this face with the input face.

See: newFaceMatch smart constructor.

Constructors

FaceMatch'

Fields

similarity :: Maybe Double

Confidence in the match of this face with the input face.

face :: Maybe Face

Describes the face properties such as the bounding box, face ID, image ID of the source image, and external image ID that you assigned.

newFaceMatch :: FaceMatch

Create a value of FaceMatch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:similarity:FaceMatch', faceMatch_similarity - Confidence in the match of this face with the input face.

$sel:face:FaceMatch', faceMatch_face - Describes the face properties such as the bounding box, face ID, image ID of the source image, and external image ID that you assigned.

faceMatch_similarity :: Lens' FaceMatch (Maybe Double)

Confidence in the match of this face with the input face.

faceMatch_face :: Lens' FaceMatch (Maybe Face)

Describes the face properties such as the bounding box, face ID, image ID of the source image, and external image ID that you assigned.