Network.AWS.Rekognition.Types.FaceSearchSettings

Description

Documentation

data FaceSearchSettings

Input face recognition parameters for an Amazon Rekognition stream processor. FaceRecognitionSettings is a request parameter for CreateStreamProcessor.

See: newFaceSearchSettings smart constructor.

Constructors

FaceSearchSettings'

Fields

collectionId :: Maybe Text

The ID of a collection that contains faces that you want to search for.

faceMatchThreshold :: Maybe Double

Minimum face match confidence score that must be met to return a result for a recognized face. Default is 80. 0 is the lowest confidence. 100 is the highest confidence.

newFaceSearchSettings :: FaceSearchSettings

Create a value of FaceSearchSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:collectionId:FaceSearchSettings', faceSearchSettings_collectionId - The ID of a collection that contains faces that you want to search for.

$sel:faceMatchThreshold:FaceSearchSettings', faceSearchSettings_faceMatchThreshold - Minimum face match confidence score that must be met to return a result for a recognized face. Default is 80. 0 is the lowest confidence. 100 is the highest confidence.

faceSearchSettings_collectionId :: Lens' FaceSearchSettings (Maybe Text)

The ID of a collection that contains faces that you want to search for.

faceSearchSettings_faceMatchThreshold :: Lens' FaceSearchSettings (Maybe Double)

Minimum face match confidence score that must be met to return a result for a recognized face. Default is 80. 0 is the lowest confidence. 100 is the highest confidence.