Network.AWS.Rekognition.Types.ProjectDescription

Description

Documentation

data ProjectDescription

A description of a Amazon Rekognition Custom Labels project.

See: newProjectDescription smart constructor.

Constructors

ProjectDescription'

Fields

status :: Maybe ProjectStatus

The current status of the project.

creationTimestamp :: Maybe POSIX

The Unix timestamp for the date and time that the project was created.

projectArn :: Maybe Text

The Amazon Resource Name (ARN) of the project.

newProjectDescription :: ProjectDescription

Create a value of ProjectDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ProjectDescription', projectDescription_status - The current status of the project.

$sel:creationTimestamp:ProjectDescription', projectDescription_creationTimestamp - The Unix timestamp for the date and time that the project was created.

$sel:projectArn:ProjectDescription', projectDescription_projectArn - The Amazon Resource Name (ARN) of the project.

projectDescription_status :: Lens' ProjectDescription (Maybe ProjectStatus)

The current status of the project.

projectDescription_creationTimestamp :: Lens' ProjectDescription (Maybe UTCTime)

The Unix timestamp for the date and time that the project was created.

projectDescription_projectArn :: Lens' ProjectDescription (Maybe Text)

The Amazon Resource Name (ARN) of the project.