Network.AWS.Rekognition.Types.ProtectiveEquipmentPerson

Description

Documentation

data ProtectiveEquipmentPerson

A person detected by a call to DetectProtectiveEquipment. The API returns all persons detected in the input image in an array of ProtectiveEquipmentPerson objects.

See: newProtectiveEquipmentPerson smart constructor.

Constructors

ProtectiveEquipmentPerson'

Fields

bodyParts :: Maybe [ProtectiveEquipmentBodyPart]

An array of body parts detected on a person's body (including body parts without PPE).

boundingBox :: Maybe BoundingBox

A bounding box around the detected person.

confidence :: Maybe Double

The confidence that Amazon Rekognition has that the bounding box contains a person.

id :: Maybe Natural

The identifier for the detected person. The identifier is only unique for a single call to DetectProtectiveEquipment.

newProtectiveEquipmentPerson :: ProtectiveEquipmentPerson

Create a value of ProtectiveEquipmentPerson with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bodyParts:ProtectiveEquipmentPerson', protectiveEquipmentPerson_bodyParts - An array of body parts detected on a person's body (including body parts without PPE).

$sel:boundingBox:ProtectiveEquipmentPerson', protectiveEquipmentPerson_boundingBox - A bounding box around the detected person.

$sel:confidence:ProtectiveEquipmentPerson', protectiveEquipmentPerson_confidence - The confidence that Amazon Rekognition has that the bounding box contains a person.

$sel:id:ProtectiveEquipmentPerson', protectiveEquipmentPerson_id - The identifier for the detected person. The identifier is only unique for a single call to DetectProtectiveEquipment.

protectiveEquipmentPerson_bodyParts :: Lens' ProtectiveEquipmentPerson (Maybe [ProtectiveEquipmentBodyPart])

An array of body parts detected on a person's body (including body parts without PPE).

protectiveEquipmentPerson_boundingBox :: Lens' ProtectiveEquipmentPerson (Maybe BoundingBox)

A bounding box around the detected person.

protectiveEquipmentPerson_confidence :: Lens' ProtectiveEquipmentPerson (Maybe Double)

The confidence that Amazon Rekognition has that the bounding box contains a person.

protectiveEquipmentPerson_id :: Lens' ProtectiveEquipmentPerson (Maybe Natural)

The identifier for the detected person. The identifier is only unique for a single call to DetectProtectiveEquipment.