Network.AWS.Rekognition.Types.ProtectiveEquipmentSummary

Description

Documentation

data ProtectiveEquipmentSummary

Summary information for required items of personal protective equipment (PPE) detected on persons by a call to DetectProtectiveEquipment. You specify the required type of PPE in the SummarizationAttributes (ProtectiveEquipmentSummarizationAttributes) input parameter. The summary includes which persons were detected wearing the required personal protective equipment (PersonsWithRequiredEquipment), which persons were detected as not wearing the required PPE (PersonsWithoutRequiredEquipment), and the persons in which a determination could not be made (PersonsIndeterminate).

To get a total for each category, use the size of the field array. For example, to find out how many people were detected as wearing the specified PPE, use the size of the PersonsWithRequiredEquipment array. If you want to find out more about a person, such as the location (BoundingBox) of the person on the image, use the person ID in each array element. Each person ID matches the ID field of a ProtectiveEquipmentPerson object returned in the Persons array by DetectProtectiveEquipment.

See: newProtectiveEquipmentSummary smart constructor.

Constructors

ProtectiveEquipmentSummary'

Fields

personsWithRequiredEquipment :: Maybe [Natural]

An array of IDs for persons who are wearing detected personal protective equipment.

personsWithoutRequiredEquipment :: Maybe [Natural]

An array of IDs for persons who are not wearing all of the types of PPE specified in the RequiredEquipmentTypes field of the detected personal protective equipment.

personsIndeterminate :: Maybe [Natural]

An array of IDs for persons where it was not possible to determine if they are wearing personal protective equipment.

newProtectiveEquipmentSummary :: ProtectiveEquipmentSummary

Create a value of ProtectiveEquipmentSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:personsWithRequiredEquipment:ProtectiveEquipmentSummary', protectiveEquipmentSummary_personsWithRequiredEquipment - An array of IDs for persons who are wearing detected personal protective equipment.

$sel:personsWithoutRequiredEquipment:ProtectiveEquipmentSummary', protectiveEquipmentSummary_personsWithoutRequiredEquipment - An array of IDs for persons who are not wearing all of the types of PPE specified in the RequiredEquipmentTypes field of the detected personal protective equipment.

$sel:personsIndeterminate:ProtectiveEquipmentSummary', protectiveEquipmentSummary_personsIndeterminate - An array of IDs for persons where it was not possible to determine if they are wearing personal protective equipment.

protectiveEquipmentSummary_personsWithRequiredEquipment :: Lens' ProtectiveEquipmentSummary (Maybe [Natural])

An array of IDs for persons who are wearing detected personal protective equipment.

protectiveEquipmentSummary_personsWithoutRequiredEquipment :: Lens' ProtectiveEquipmentSummary (Maybe [Natural])

An array of IDs for persons who are not wearing all of the types of PPE specified in the RequiredEquipmentTypes field of the detected personal protective equipment.

protectiveEquipmentSummary_personsIndeterminate :: Lens' ProtectiveEquipmentSummary (Maybe [Natural])

An array of IDs for persons where it was not possible to determine if they are wearing personal protective equipment.