Network.AWS.Rekognition.Types.TestingData

Description

Documentation

data TestingData

The dataset used for testing. Optionally, if AutoCreate is set, Amazon Rekognition Custom Labels creates a testing dataset using an 80/20 split of the training dataset.

See: newTestingData smart constructor.

Constructors

TestingData'

Fields

autoCreate :: Maybe Bool

If specified, Amazon Rekognition Custom Labels creates a testing dataset with an 80/20 split of the training dataset.

assets :: Maybe [Asset]

The assets used for testing.

newTestingData :: TestingData

Create a value of TestingData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoCreate:TestingData', testingData_autoCreate - If specified, Amazon Rekognition Custom Labels creates a testing dataset with an 80/20 split of the training dataset.

$sel:assets:TestingData', testingData_assets - The assets used for testing.

testingData_autoCreate :: Lens' TestingData (Maybe Bool)

If specified, Amazon Rekognition Custom Labels creates a testing dataset with an 80/20 split of the training dataset.

testingData_assets :: Lens' TestingData (Maybe [Asset])

The assets used for testing.