Network.AWS.Rekognition.Types.TextDetection

Description

Documentation

data TextDetection

Information about a word or line of text detected by DetectText.

The DetectedText field contains the text that Amazon Rekognition detected in the image.

Every word and line has an identifier (Id). Each word belongs to a line and has a parent identifier (ParentId) that identifies the line of text in which the word appears. The word Id is also an index for the word within a line of words.

For more information, see Detecting Text in the Amazon Rekognition Developer Guide.

See: newTextDetection smart constructor.

Constructors

TextDetection'

Fields

detectedText :: Maybe Text

The word or line of text recognized by Amazon Rekognition.

confidence :: Maybe Double

The confidence that Amazon Rekognition has in the accuracy of the detected text and the accuracy of the geometry points around the detected text.

geometry :: Maybe Geometry

The location of the detected text on the image. Includes an axis aligned coarse bounding box surrounding the text and a finer grain polygon for more accurate spatial information.

id :: Maybe Natural

The identifier for the detected text. The identifier is only unique for a single call to DetectText.

type' :: Maybe TextTypes

The type of text that was detected.

parentId :: Maybe Natural

The Parent identifier for the detected text identified by the value of ID. If the type of detected text is LINE, the value of ParentId is Null.

newTextDetection :: TextDetection

Create a value of TextDetection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detectedText:TextDetection', textDetection_detectedText - The word or line of text recognized by Amazon Rekognition.

$sel:confidence:TextDetection', textDetection_confidence - The confidence that Amazon Rekognition has in the accuracy of the detected text and the accuracy of the geometry points around the detected text.

$sel:geometry:TextDetection', textDetection_geometry - The location of the detected text on the image. Includes an axis aligned coarse bounding box surrounding the text and a finer grain polygon for more accurate spatial information.

$sel:id:TextDetection', textDetection_id - The identifier for the detected text. The identifier is only unique for a single call to DetectText.

$sel:type':TextDetection', textDetection_type - The type of text that was detected.

$sel:parentId:TextDetection', textDetection_parentId - The Parent identifier for the detected text identified by the value of ID. If the type of detected text is LINE, the value of ParentId is Null.

textDetection_detectedText :: Lens' TextDetection (Maybe Text)

The word or line of text recognized by Amazon Rekognition.

textDetection_confidence :: Lens' TextDetection (Maybe Double)

The confidence that Amazon Rekognition has in the accuracy of the detected text and the accuracy of the geometry points around the detected text.

textDetection_geometry :: Lens' TextDetection (Maybe Geometry)

The location of the detected text on the image. Includes an axis aligned coarse bounding box surrounding the text and a finer grain polygon for more accurate spatial information.

textDetection_id :: Lens' TextDetection (Maybe Natural)

The identifier for the detected text. The identifier is only unique for a single call to DetectText.

textDetection_type :: Lens' TextDetection (Maybe TextTypes)

The type of text that was detected.

textDetection_parentId :: Lens' TextDetection (Maybe Natural)

The Parent identifier for the detected text identified by the value of ID. If the type of detected text is LINE, the value of ParentId is Null.