Network.AWS.ResourceGroups

Description

Derived from API version 2017-11-27 of the AWS service descriptions, licensed under Apache 2.0.

AWS Resource Groups

AWS Resource Groups lets you organize AWS resources such as Amazon EC2 instances, Amazon Relational Database Service databases, and Amazon S3 buckets into groups using criteria that you define as tags. A resource group is a collection of resources that match the resource types specified in a query, and share one or more tags or portions of tags. You can create a group of resources based on their roles in your cloud infrastructure, lifecycle stages, regions, application layers, or virtually any criteria. Resource Groups enable you to automate management tasks, such as those in AWS Systems Manager Automation documents, on tag-related resources in AWS Systems Manager. Groups of tagged resources also let you quickly view a custom console in AWS Systems Manager that shows AWS Config compliance and other monitoring data about member resources.

To create a resource group, build a resource query, and specify tags that identify the criteria that members of the group have in common. Tags are key-value pairs.

For more information about Resource Groups, see the AWS Resource Groups User Guide.

AWS Resource Groups uses a REST-compliant API that you can use to perform the following types of operations.

Create, Read, Update, and Delete (CRUD) operations on resource groups and resource query entities

Applying, editing, and removing tags from resource groups

Resolving resource group member ARNs so they can be returned as search results

Getting data about resources that are members of a group

Searching AWS resources based on a resource query

Service Configuration

defaultService :: Service

API version 2017-11-27 of the Amazon Resource Groups SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by ResourceGroups.

NotFoundException

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

One or more of the specified resources don't exist.

BadRequestException

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request includes one or more parameters that violate validation rules.

UnauthorizedException

_UnauthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because it doesn't have valid credentials for the target resource.

InternalServerErrorException

_InternalServerErrorException :: AsError a => Getting (First ServiceError) a ServiceError

An internal error occurred while processing the request. Try again later.

ForbiddenException

_ForbiddenException :: AsError a => Getting (First ServiceError) a ServiceError

The caller isn't authorized to make the request. Check permissions.

MethodNotAllowedException

_MethodNotAllowedException :: AsError a => Getting (First ServiceError) a ServiceError

The request uses an HTTP method that isn't allowed for the specified resource.

TooManyRequestsException

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

You've exceeded throttling limits by making too many requests in a period of time.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetGroupConfiguration

data GetGroupConfiguration

See: newGetGroupConfiguration smart constructor.

Constructors

GetGroupConfiguration' (Maybe Text)

newGetGroupConfiguration :: GetGroupConfiguration

Create a value of GetGroupConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:group':GetGroupConfiguration', getGroupConfiguration_group - The name or the ARN of the resource group.

data GetGroupConfigurationResponse

See: newGetGroupConfigurationResponse smart constructor.

Constructors

GetGroupConfigurationResponse' (Maybe GroupConfiguration) Int

newGetGroupConfigurationResponse

Arguments

:: Int
$sel:httpStatus:GetGroupConfigurationResponse'

-> GetGroupConfigurationResponse

Create a value of GetGroupConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupConfiguration:GetGroupConfigurationResponse', getGroupConfigurationResponse_groupConfiguration - The service configuration associated with the specified group. For details about the service configuration syntax, see Service configurations for resource groups.

$sel:httpStatus:GetGroupConfigurationResponse', getGroupConfigurationResponse_httpStatus - The response's http status code.

PutGroupConfiguration

data PutGroupConfiguration

See: newPutGroupConfiguration smart constructor.

Constructors

PutGroupConfiguration' (Maybe [GroupConfigurationItem]) (Maybe Text)

newPutGroupConfiguration :: PutGroupConfiguration

Create a value of PutGroupConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configuration:PutGroupConfiguration', putGroupConfiguration_configuration - The new configuration to associate with the specified group. A configuration associates the resource group with an AWS service and specifies how the service can interact with the resources in the group. A configuration is an array of GroupConfigurationItem elements.

For information about the syntax of a service configuration, see Service configurations for resource groups.

A resource group can contain either a Configuration or a ResourceQuery, but not both.

$sel:group':PutGroupConfiguration', putGroupConfiguration_group - The name or ARN of the resource group with the configuration that you want to update.

data PutGroupConfigurationResponse

See: newPutGroupConfigurationResponse smart constructor.

Constructors

PutGroupConfigurationResponse' Int

newPutGroupConfigurationResponse

Arguments

:: Int
$sel:httpStatus:PutGroupConfigurationResponse'

-> PutGroupConfigurationResponse

Create a value of PutGroupConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutGroupConfigurationResponse', putGroupConfigurationResponse_httpStatus - The response's http status code.

ListGroups (Paginated)

data ListGroups

See: newListGroups smart constructor.

Constructors

ListGroups' (Maybe Text) (Maybe Natural) (Maybe [GroupFilter])

newListGroups :: ListGroups

Create a value of ListGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListGroups', listGroups_nextToken - The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value provided by a previous call's NextToken response to indicate where the output should continue from.

$sel:maxResults:ListGroups', listGroups_maxResults - The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that the service might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

$sel:filters:ListGroups', listGroups_filters - Filters, formatted as GroupFilter objects, that you want to apply to a ListGroups operation.

resource-type - Filter the results to include only those of the specified resource types. Specify up to five resource types in the format AWS::ServiceCode::ResourceType . For example, AWS::EC2::Instance, or AWS::S3::Bucket.

configuration-type - Filter the results to include only those groups that have the specified configuration types attached. The current supported values are:

AWS:EC2::CapacityReservationPool

AWS:EC2::HostManagement

data ListGroupsResponse

See: newListGroupsResponse smart constructor.

Constructors

ListGroupsResponse' (Maybe [Group]) (Maybe Text) (Maybe [GroupIdentifier]) Int

newListGroupsResponse

Arguments

:: Int
$sel:httpStatus:ListGroupsResponse'

-> ListGroupsResponse

Create a value of ListGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groups:ListGroupsResponse', listGroupsResponse_groups - /__Deprecated - don't use this field. Use the GroupIdentifiers response field instead.__/

$sel:nextToken:ListGroups', listGroupsResponse_nextToken - If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

$sel:groupIdentifiers:ListGroupsResponse', listGroupsResponse_groupIdentifiers - A list of GroupIdentifier objects. Each identifier is an object that contains both the Name and the GroupArn.

$sel:httpStatus:ListGroupsResponse', listGroupsResponse_httpStatus - The response's http status code.

CreateGroup

data CreateGroup

See: newCreateGroup smart constructor.

Constructors

CreateGroup' (Maybe [GroupConfigurationItem]) (Maybe (HashMap Text Text)) (Maybe Text) (Maybe ResourceQuery) Text

newCreateGroup

Arguments

:: Text
$sel:name:CreateGroup'

-> CreateGroup

Create a value of CreateGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configuration:CreateGroup', createGroup_configuration - A configuration associates the resource group with an AWS service and specifies how the service can interact with the resources in the group. A configuration is an array of GroupConfigurationItem elements. For details about the syntax of service configurations, see Service configurations for resource groups.

A resource group can contain either a Configuration or a ResourceQuery, but not both.

$sel:tags:CreateGroup', createGroup_tags - The tags to add to the group. A tag is key-value pair string.

$sel:description:CreateGroup', createGroup_description - The description of the resource group. Descriptions can consist of letters, numbers, hyphens, underscores, periods, and spaces.

$sel:resourceQuery:CreateGroup', createGroup_resourceQuery - The resource query that determines which AWS resources are members of this group. For more information about resource queries, see Create a tag-based group in Resource Groups.

A resource group can contain either a ResourceQuery or a Configuration, but not both.

$sel:name:CreateGroup', createGroup_name - The name of the group, which is the identifier of the group in other operations. You can't change the name of a resource group after you create it. A resource group name can consist of letters, numbers, hyphens, periods, and underscores. The name cannot start with AWS or aws; these are reserved. A resource group name must be unique within each AWS Region in your AWS account.

data CreateGroupResponse

See: newCreateGroupResponse smart constructor.

Constructors

CreateGroupResponse' (Maybe GroupConfiguration) (Maybe Group) (Maybe (HashMap Text Text)) (Maybe ResourceQuery) Int

newCreateGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateGroupResponse'

-> CreateGroupResponse

Create a value of CreateGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupConfiguration:CreateGroupResponse', createGroupResponse_groupConfiguration - The service configuration associated with the resource group. For details about the syntax of a service configuration, see Service configurations for resource groups.

$sel:group':CreateGroupResponse', createGroupResponse_group - The description of the resource group.

$sel:tags:CreateGroup', createGroupResponse_tags - The tags associated with the group.

$sel:resourceQuery:CreateGroup', createGroupResponse_resourceQuery - The resource query associated with the group. For more information about resource queries, see Create a tag-based group in Resource Groups.

$sel:httpStatus:CreateGroupResponse', createGroupResponse_httpStatus - The response's http status code.

GetGroupQuery

data GetGroupQuery

See: newGetGroupQuery smart constructor.

Constructors

GetGroupQuery' (Maybe Text) (Maybe Text)

newGetGroupQuery :: GetGroupQuery

Create a value of GetGroupQuery with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupName:GetGroupQuery', getGroupQuery_groupName - Don't use this parameter. Use Group instead.

$sel:group':GetGroupQuery', getGroupQuery_group - The name or the ARN of the resource group to query.

data GetGroupQueryResponse

See: newGetGroupQueryResponse smart constructor.

Constructors

GetGroupQueryResponse' (Maybe GroupQuery) Int

newGetGroupQueryResponse

Arguments

:: Int
$sel:httpStatus:GetGroupQueryResponse'

-> GetGroupQueryResponse

Create a value of GetGroupQueryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupQuery:GetGroupQueryResponse', getGroupQueryResponse_groupQuery - The resource query associated with the specified group. For more information about resource queries, see Create a tag-based group in Resource Groups.

$sel:httpStatus:GetGroupQueryResponse', getGroupQueryResponse_httpStatus - The response's http status code.

SearchResources (Paginated)

data SearchResources

See: newSearchResources smart constructor.

Constructors

SearchResources' (Maybe Text) (Maybe Natural) ResourceQuery

newSearchResources

Arguments

:: ResourceQuery
$sel:resourceQuery:SearchResources'

-> SearchResources

Create a value of SearchResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchResources', searchResources_nextToken - The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value provided by a previous call's NextToken response to indicate where the output should continue from.

$sel:maxResults:SearchResources', searchResources_maxResults - The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that the service might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

$sel:resourceQuery:SearchResources', searchResources_resourceQuery - The search query, using the same formats that are supported for resource group definition. For more information, see CreateGroup.

data SearchResourcesResponse

See: newSearchResourcesResponse smart constructor.

Constructors

SearchResourcesResponse' (Maybe Text) (Maybe [ResourceIdentifier]) (Maybe [QueryError]) Int

newSearchResourcesResponse

Arguments

:: Int
$sel:httpStatus:SearchResourcesResponse'

-> SearchResourcesResponse

Create a value of SearchResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchResources', searchResourcesResponse_nextToken - If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

$sel:resourceIdentifiers:SearchResourcesResponse', searchResourcesResponse_resourceIdentifiers - The ARNs and resource types of resources that are members of the group that you specified.

$sel:queryErrors:SearchResourcesResponse', searchResourcesResponse_queryErrors - A list of QueryError objects. Each error is an object that contains ErrorCode and Message structures. Possible values for ErrorCode are CLOUDFORMATION_STACK_INACTIVE and CLOUDFORMATION_STACK_NOT_EXISTING.

$sel:httpStatus:SearchResourcesResponse', searchResourcesResponse_httpStatus - The response's http status code.

GetTags

data GetTags

See: newGetTags smart constructor.

Constructors

GetTags' Text

newGetTags

Arguments

:: Text
$sel:arn:GetTags'

-> GetTags

Create a value of GetTags with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:GetTags', getTags_arn - The ARN of the resource group whose tags you want to retrieve.

data GetTagsResponse

See: newGetTagsResponse smart constructor.

Constructors

GetTagsResponse' (Maybe Text) (Maybe (HashMap Text Text)) Int

newGetTagsResponse

Arguments

:: Int
$sel:httpStatus:GetTagsResponse'

-> GetTagsResponse

Create a value of GetTagsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:GetTags', getTagsResponse_arn - The ARN of the tagged resource group.

$sel:tags:GetTagsResponse', getTagsResponse_tags - The tags associated with the specified resource group.

$sel:httpStatus:GetTagsResponse', getTagsResponse_httpStatus - The response's http status code.

UpdateGroupQuery

data UpdateGroupQuery

See: newUpdateGroupQuery smart constructor.

Constructors

UpdateGroupQuery' (Maybe Text) (Maybe Text) ResourceQuery

newUpdateGroupQuery

Arguments

:: ResourceQuery
$sel:resourceQuery:UpdateGroupQuery'

-> UpdateGroupQuery

Create a value of UpdateGroupQuery with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupName:UpdateGroupQuery', updateGroupQuery_groupName - Don't use this parameter. Use Group instead.

$sel:group':UpdateGroupQuery', updateGroupQuery_group - The name or the ARN of the resource group to query.

$sel:resourceQuery:UpdateGroupQuery', updateGroupQuery_resourceQuery - The resource query to determine which AWS resources are members of this resource group.

A resource group can contain either a Configuration or a ResourceQuery, but not both.

data UpdateGroupQueryResponse

See: newUpdateGroupQueryResponse smart constructor.

Constructors

UpdateGroupQueryResponse' (Maybe GroupQuery) Int

newUpdateGroupQueryResponse

Arguments

:: Int
$sel:httpStatus:UpdateGroupQueryResponse'

-> UpdateGroupQueryResponse

Create a value of UpdateGroupQueryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupQuery:UpdateGroupQueryResponse', updateGroupQueryResponse_groupQuery - The updated resource query associated with the resource group after the update.

$sel:httpStatus:UpdateGroupQueryResponse', updateGroupQueryResponse_httpStatus - The response's http status code.

ListGroupResources (Paginated)

data ListGroupResources

See: newListGroupResources smart constructor.

Constructors

ListGroupResources' (Maybe Text) (Maybe Natural) (Maybe Text) (Maybe Text) (Maybe [ResourceFilter])

newListGroupResources :: ListGroupResources

Create a value of ListGroupResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListGroupResources', listGroupResources_nextToken - The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value provided by a previous call's NextToken response to indicate where the output should continue from.

$sel:maxResults:ListGroupResources', listGroupResources_maxResults - The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that the service might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

$sel:groupName:ListGroupResources', listGroupResources_groupName - /__Deprecated - don't use this parameter. Use the Group request field instead.__/

$sel:group':ListGroupResources', listGroupResources_group - The name or the ARN of the resource group

$sel:filters:ListGroupResources', listGroupResources_filters - Filters, formatted as ResourceFilter objects, that you want to apply to a ListGroupResources operation. Filters the results to include only those of the specified resource types.

resource-type - Filter resources by their type. Specify up to five resource types in the format AWS::ServiceCode::ResourceType. For example, AWS::EC2::Instance, or AWS::S3::Bucket.

When you specify a resource-type filter for ListGroupResources, AWS Resource Groups validates your filter resource types against the types that are defined in the query associated with the group. For example, if a group contains only S3 buckets because its query specifies only that resource type, but your resource-type filter includes EC2 instances, AWS Resource Groups does not filter for EC2 instances. In this case, a ListGroupResources request returns a BadRequestException error with a message similar to the following:

The resource types specified as filters in the request are not valid.

The error includes a list of resource types that failed the validation because they are not part of the query associated with the group. This validation doesn't occur when the group query specifies AWS::AllSupported, because a group based on such a query can contain any of the allowed resource types for the query type (tag-based or AWS CloudFormation stack-based queries).

data ListGroupResourcesResponse

See: newListGroupResourcesResponse smart constructor.

Constructors

ListGroupResourcesResponse' (Maybe Text) (Maybe [ResourceIdentifier]) (Maybe [QueryError]) (Maybe [ListGroupResourcesItem]) Int

newListGroupResourcesResponse

Arguments

:: Int
$sel:httpStatus:ListGroupResourcesResponse'

-> ListGroupResourcesResponse

Create a value of ListGroupResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListGroupResources', listGroupResourcesResponse_nextToken - If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

$sel:resourceIdentifiers:ListGroupResourcesResponse', listGroupResourcesResponse_resourceIdentifiers - __/Deprecated - don't use this parameter. Use the Resources response field instead./__

$sel:queryErrors:ListGroupResourcesResponse', listGroupResourcesResponse_queryErrors - A list of QueryError objects. Each error is an object that contains ErrorCode and Message structures. Possible values for ErrorCode are CLOUDFORMATION_STACK_INACTIVE and CLOUDFORMATION_STACK_NOT_EXISTING.

$sel:resources:ListGroupResourcesResponse', listGroupResourcesResponse_resources - An array of resources from which you can determine each resource's identity, type, and group membership status.

$sel:httpStatus:ListGroupResourcesResponse', listGroupResourcesResponse_httpStatus - The response's http status code.

Untag

data Untag

See: newUntag smart constructor.

Constructors

Untag' Text [Text]

newUntag

Arguments

:: Text
$sel:arn:Untag'

-> Untag

Create a value of Untag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Untag', untag_arn - The ARN of the resource group from which to remove tags. The command removed both the specified keys and any values associated with those keys.

$sel:keys:Untag', untag_keys - The keys of the tags to be removed.

data UntagResponse

See: newUntagResponse smart constructor.

Constructors

UntagResponse' (Maybe Text) (Maybe [Text]) Int

newUntagResponse

Arguments

:: Int
$sel:httpStatus:UntagResponse'

-> UntagResponse

Create a value of UntagResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Untag', untagResponse_arn - The ARN of the resource group from which tags have been removed.

$sel:keys:Untag', untagResponse_keys - The keys of the tags that were removed.

$sel:httpStatus:UntagResponse', untagResponse_httpStatus - The response's http status code.

GetGroup

data GetGroup

See: newGetGroup smart constructor.

Constructors

GetGroup' (Maybe Text) (Maybe Text)

newGetGroup :: GetGroup

Create a value of GetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupName:GetGroup', getGroup_groupName - Deprecated - don't use this parameter. Use Group instead.

$sel:group':GetGroup', getGroup_group - The name or the ARN of the resource group to retrieve.

data GetGroupResponse

See: newGetGroupResponse smart constructor.

Constructors

GetGroupResponse' (Maybe Group) Int

newGetGroupResponse

Arguments

:: Int
$sel:httpStatus:GetGroupResponse'

-> GetGroupResponse

Create a value of GetGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:group':GetGroup', getGroupResponse_group - A full description of the resource group.

$sel:httpStatus:GetGroupResponse', getGroupResponse_httpStatus - The response's http status code.

DeleteGroup

data DeleteGroup

See: newDeleteGroup smart constructor.

Constructors

DeleteGroup' (Maybe Text) (Maybe Text)

newDeleteGroup :: DeleteGroup

Create a value of DeleteGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupName:DeleteGroup', deleteGroup_groupName - Deprecated - don't use this parameter. Use Group instead.

$sel:group':DeleteGroup', deleteGroup_group - The name or the ARN of the resource group to delete.

data DeleteGroupResponse

See: newDeleteGroupResponse smart constructor.

Constructors

DeleteGroupResponse' (Maybe Group) Int

newDeleteGroupResponse

Arguments

:: Int
$sel:httpStatus:DeleteGroupResponse'

-> DeleteGroupResponse

Create a value of DeleteGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:group':DeleteGroup', deleteGroupResponse_group - A full description of the deleted resource group.

$sel:httpStatus:DeleteGroupResponse', deleteGroupResponse_httpStatus - The response's http status code.

GroupResources

data GroupResources

See: newGroupResources smart constructor.

Constructors

GroupResources' Text (NonEmpty Text)

newGroupResources

Arguments

:: Text
$sel:group':GroupResources'

-> NonEmpty Text
$sel:resourceArns:GroupResources'

-> GroupResources

Create a value of GroupResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:group':GroupResources', groupResources_group - The name or the ARN of the resource group to add resources to.

$sel:resourceArns:GroupResources', groupResources_resourceArns - The list of ARNs for resources to be added to the group.

data GroupResourcesResponse

See: newGroupResourcesResponse smart constructor.

Constructors

GroupResourcesResponse' (Maybe (NonEmpty Text)) (Maybe [PendingResource]) (Maybe [FailedResource]) Int

newGroupResourcesResponse

Arguments

:: Int
$sel:httpStatus:GroupResourcesResponse'

-> GroupResourcesResponse

Create a value of GroupResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:succeeded:GroupResourcesResponse', groupResourcesResponse_succeeded - A list of ARNs of resources that were successfully added to the group by this operation.

$sel:pending:GroupResourcesResponse', groupResourcesResponse_pending - A list of ARNs of any resources that are still in the process of being added to the group by this operation. These pending additions continue asynchronously. You can check the status of pending additions by using the ListGroupResources operation, and checking the Resources array in the response and the Status field of each object in that array.

$sel:failed:GroupResourcesResponse', groupResourcesResponse_failed - A list of ARNs of any resources that failed to be added to the group by this operation.

$sel:httpStatus:GroupResourcesResponse', groupResourcesResponse_httpStatus - The response's http status code.

UpdateGroup

data UpdateGroup

See: newUpdateGroup smart constructor.

Constructors

UpdateGroup' (Maybe Text) (Maybe Text) (Maybe Text)

newUpdateGroup :: UpdateGroup

Create a value of UpdateGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupName:UpdateGroup', updateGroup_groupName - Don't use this parameter. Use Group instead.

$sel:group':UpdateGroup', updateGroup_group - The name or the ARN of the resource group to modify.

$sel:description:UpdateGroup', updateGroup_description - The new description that you want to update the resource group with. Descriptions can contain letters, numbers, hyphens, underscores, periods, and spaces.

data UpdateGroupResponse

See: newUpdateGroupResponse smart constructor.

Constructors

UpdateGroupResponse' (Maybe Group) Int

newUpdateGroupResponse

Arguments

:: Int
$sel:httpStatus:UpdateGroupResponse'

-> UpdateGroupResponse

Create a value of UpdateGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:group':UpdateGroup', updateGroupResponse_group - The update description of the resource group.

$sel:httpStatus:UpdateGroupResponse', updateGroupResponse_httpStatus - The response's http status code.

UngroupResources

data UngroupResources

See: newUngroupResources smart constructor.

Constructors

UngroupResources' Text (NonEmpty Text)

newUngroupResources

Arguments

:: Text
$sel:group':UngroupResources'

-> NonEmpty Text
$sel:resourceArns:UngroupResources'

-> UngroupResources

Create a value of UngroupResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:group':UngroupResources', ungroupResources_group - The name or the ARN of the resource group from which to remove the resources.

$sel:resourceArns:UngroupResources', ungroupResources_resourceArns - The ARNs of the resources to be removed from the group.

data UngroupResourcesResponse

See: newUngroupResourcesResponse smart constructor.

Constructors

UngroupResourcesResponse' (Maybe (NonEmpty Text)) (Maybe [PendingResource]) (Maybe [FailedResource]) Int

newUngroupResourcesResponse

Arguments

:: Int
$sel:httpStatus:UngroupResourcesResponse'

-> UngroupResourcesResponse

Create a value of UngroupResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:succeeded:UngroupResourcesResponse', ungroupResourcesResponse_succeeded - A list of resources that were successfully removed from the group by this operation.

$sel:pending:UngroupResourcesResponse', ungroupResourcesResponse_pending - A list of any resources that are still in the process of being removed from the group by this operation. These pending removals continue asynchronously. You can check the status of pending removals by using the ListGroupResources operation. After the resource is successfully removed, it no longer appears in the response.

$sel:failed:UngroupResourcesResponse', ungroupResourcesResponse_failed - A list of any resources that failed to be removed from the group by this operation.

$sel:httpStatus:UngroupResourcesResponse', ungroupResourcesResponse_httpStatus - The response's http status code.

Tag

data Tag

See: newTag smart constructor.

Constructors

Tag' Text (HashMap Text Text)

newTag

Arguments

:: Text
$sel:arn:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Tag', tag_arn - The ARN of the resource group to which to add tags.

$sel:tags:Tag', tag_tags - The tags to add to the specified resource group. A tag is a string-to-string map of key-value pairs.

data TagResponse

See: newTagResponse smart constructor.

Constructors

TagResponse' (Maybe Text) (Maybe (HashMap Text Text)) Int

newTagResponse

Arguments

:: Int
$sel:httpStatus:TagResponse'

-> TagResponse

Create a value of TagResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Tag', tagResponse_arn - The ARN of the tagged resource.

$sel:tags:Tag', tagResponse_tags - The tags that have been added to the specified resource group.

$sel:httpStatus:TagResponse', tagResponse_httpStatus - The response's http status code.

Types

GroupConfigurationStatus

newtype GroupConfigurationStatus

Constructors

GroupConfigurationStatus'

Fields

fromGroupConfigurationStatus :: Text

Bundled Patterns

pattern GroupConfigurationStatus_UPDATE_COMPLETE :: GroupConfigurationStatus

pattern GroupConfigurationStatus_UPDATE_FAILED :: GroupConfigurationStatus

pattern GroupConfigurationStatus_UPDATING :: GroupConfigurationStatus

GroupFilterName

newtype GroupFilterName

Constructors

GroupFilterName'

Fields

fromGroupFilterName :: Text

Bundled Patterns

pattern GroupFilterName_Configuration_type :: GroupFilterName

pattern GroupFilterName_Resource_type :: GroupFilterName

QueryErrorCode

newtype QueryErrorCode

Constructors

QueryErrorCode'

Fields

fromQueryErrorCode :: Text

Bundled Patterns

pattern QueryErrorCode_CLOUDFORMATION_STACK_INACTIVE :: QueryErrorCode

pattern QueryErrorCode_CLOUDFORMATION_STACK_NOT_EXISTING :: QueryErrorCode

QueryType

newtype QueryType

Constructors

QueryType'

Fields

fromQueryType :: Text

Bundled Patterns

pattern QueryType_CLOUDFORMATION_STACK_1_0 :: QueryType

pattern QueryType_TAG_FILTERS_1_0 :: QueryType

ResourceFilterName

newtype ResourceFilterName

Constructors

ResourceFilterName'

Fields

fromResourceFilterName :: Text

Bundled Patterns

pattern ResourceFilterName_Resource_type :: ResourceFilterName

ResourceStatusValue

newtype ResourceStatusValue

Constructors

ResourceStatusValue'

Fields

fromResourceStatusValue :: Text

Bundled Patterns

pattern ResourceStatusValue_PENDING :: ResourceStatusValue

FailedResource

data FailedResource

A resource that failed to be added to or removed from a group.

See: newFailedResource smart constructor.

Constructors

FailedResource' (Maybe Text) (Maybe Text) (Maybe Text)

newFailedResource :: FailedResource

Create a value of FailedResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:FailedResource', failedResource_resourceArn - The ARN of the resource that failed to be added or removed.

$sel:errorMessage:FailedResource', failedResource_errorMessage - The error message text associated with the failure.

$sel:errorCode:FailedResource', failedResource_errorCode - The error code associated with the failure.

Group

data Group

A resource group that contains AWS resources. You can assign resources to the group by associating either of the following elements with the group:

ResourceQuery - Use a resource query to specify a set of tag keys and values. All resources in the same AWS Region and AWS account that have those keys with the same values are included in the group. You can add a resource query when you create the group, or later by using the PutGroupConfiguration operation.

GroupConfiguration - Use a service configuration to associate the group with an AWS service. The configuration specifies which resource types can be included in the group.

See: newGroup smart constructor.

Constructors

Group' (Maybe Text) Text Text

newGroup

Arguments

:: Text
$sel:groupArn:Group'

-> Text
$sel:name:Group'

-> Group

Create a value of Group with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:Group', group_description - The description of the resource group.

$sel:groupArn:Group', group_groupArn - The ARN of the resource group.

$sel:name:Group', group_name - The name of the resource group.

GroupConfiguration

data GroupConfiguration

A service configuration associated with a resource group. The configuration options are determined by the AWS service that defines the Type, and specifies which resources can be included in the group. You can add a service configuration when you create the group by using CreateGroup, or later by using the PutGroupConfiguration operation. For details about group service configuration syntax, see Service configurations for resource groups.

See: newGroupConfiguration smart constructor.

Constructors

GroupConfiguration' (Maybe GroupConfigurationStatus) (Maybe [GroupConfigurationItem]) (Maybe Text) (Maybe [GroupConfigurationItem])

newGroupConfiguration :: GroupConfiguration

Create a value of GroupConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GroupConfiguration', groupConfiguration_status - The current status of an attempt to update the group configuration.

$sel:configuration:GroupConfiguration', groupConfiguration_configuration - The configuration currently associated with the group and in effect.

$sel:failureReason:GroupConfiguration', groupConfiguration_failureReason - If present, the reason why a request to update the group configuration failed.

$sel:proposedConfiguration:GroupConfiguration', groupConfiguration_proposedConfiguration - If present, the new configuration that is in the process of being applied to the group.

GroupConfigurationItem

data GroupConfigurationItem

An item in a group configuration. A group service configuration can have one or more items. For details about group service configuration syntax, see Service configurations for resource groups.

See: newGroupConfigurationItem smart constructor.

Constructors

GroupConfigurationItem' (Maybe [GroupConfigurationParameter]) Text

newGroupConfigurationItem

Arguments

:: Text
$sel:type':GroupConfigurationItem'

-> GroupConfigurationItem

Create a value of GroupConfigurationItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameters:GroupConfigurationItem', groupConfigurationItem_parameters - A collection of parameters for this group configuration item. For the list of parameters that you can use with each configuration item type, see Supported resource types and parameters.

$sel:type':GroupConfigurationItem', groupConfigurationItem_type - Specifies the type of group configuration item. Each item must have a unique value for type. For the list of types that you can specify for a configuration item, see Supported resource types and parameters.

GroupConfigurationParameter

data GroupConfigurationParameter

A parameter for a group configuration item. For details about group service configuration syntax, see Service configurations for resource groups.

See: newGroupConfigurationParameter smart constructor.

Constructors

GroupConfigurationParameter' (Maybe [Text]) Text

newGroupConfigurationParameter

Arguments

:: Text
$sel:name:GroupConfigurationParameter'

-> GroupConfigurationParameter

Create a value of GroupConfigurationParameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:GroupConfigurationParameter', groupConfigurationParameter_values - The value or values to be used for the specified parameter. For the list of values you can use with each parameter, see Supported resource types and parameters.

$sel:name:GroupConfigurationParameter', groupConfigurationParameter_name - The name of the group configuration parameter. For the list of parameters that you can use with each configuration item type, see Supported resource types and parameters.

GroupFilter

data GroupFilter

A filter collection that you can use to restrict the results from a List operation to only those you want to include.

See: newGroupFilter smart constructor.

Constructors

GroupFilter' GroupFilterName (NonEmpty Text)

newGroupFilter

Arguments

:: GroupFilterName
$sel:name:GroupFilter'

-> NonEmpty Text
$sel:values:GroupFilter'

-> GroupFilter

Create a value of GroupFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GroupFilter', groupFilter_name - The name of the filter. Filter names are case-sensitive.

$sel:values:GroupFilter', groupFilter_values - One or more filter values. Allowed filter values vary by group filter name, and are case-sensitive.

GroupIdentifier

data GroupIdentifier

The unique identifiers for a resource group.

See: newGroupIdentifier smart constructor.

Constructors

GroupIdentifier' (Maybe Text) (Maybe Text)

newGroupIdentifier :: GroupIdentifier

Create a value of GroupIdentifier with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupName:GroupIdentifier', groupIdentifier_groupName - The name of the resource group.

$sel:groupArn:GroupIdentifier', groupIdentifier_groupArn - The ARN of the resource group.

GroupQuery

data GroupQuery

A mapping of a query attached to a resource group that determines the AWS resources that are members of the group.

See: newGroupQuery smart constructor.

Constructors

GroupQuery' Text ResourceQuery

newGroupQuery

Arguments

:: Text
$sel:groupName:GroupQuery'

-> ResourceQuery
$sel:resourceQuery:GroupQuery'

-> GroupQuery

Create a value of GroupQuery with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupName:GroupQuery', groupQuery_groupName - The name of the resource group that is associated with the specified resource query.

$sel:resourceQuery:GroupQuery', groupQuery_resourceQuery - The resource query that determines which AWS resources are members of the associated resource group.

ListGroupResourcesItem

data ListGroupResourcesItem

A structure returned by the ListGroupResources operation that contains identity and group membership status information for one of the resources in the group.

See: newListGroupResourcesItem smart constructor.

Constructors

ListGroupResourcesItem' (Maybe ResourceStatus) (Maybe ResourceIdentifier)

newListGroupResourcesItem :: ListGroupResourcesItem

Create a value of ListGroupResourcesItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListGroupResourcesItem', listGroupResourcesItem_status - A structure that contains the status of this resource's membership in the group.

This field is present in the response only if the group is of type AWS::EC2::HostManagement.

$sel:identifier:ListGroupResourcesItem', listGroupResourcesItem_identifier - Undocumented member.

PendingResource

data PendingResource

A structure that identifies a resource that is currently pending addition to the group as a member. Adding a resource to a resource group happens asynchronously as a background task and this one isn't completed yet.

See: newPendingResource smart constructor.

Constructors

PendingResource' (Maybe Text)

newPendingResource :: PendingResource

Create a value of PendingResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:PendingResource', pendingResource_resourceArn - The Amazon resource name (ARN) of the resource that's in a pending state.

QueryError

data QueryError

A two-part error structure that can occur in ListGroupResources or SearchResources operations on CloudFormation stack-based queries. The error occurs if the CloudFormation stack on which the query is based either does not exist, or has a status that renders the stack inactive. A QueryError occurrence does not necessarily mean that AWS Resource Groups could not complete the operation, but the resulting group might have no member resources.

See: newQueryError smart constructor.

Constructors

QueryError' (Maybe Text) (Maybe QueryErrorCode)

newQueryError :: QueryError

Create a value of QueryError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:QueryError', queryError_message - A message that explains the ErrorCode value. Messages might state that the specified CloudFormation stack does not exist (or no longer exists). For CLOUDFORMATION_STACK_INACTIVE, the message typically states that the CloudFormation stack has a status that is not (or no longer) active, such as CREATE_FAILED.

$sel:errorCode:QueryError', queryError_errorCode - Possible values are CLOUDFORMATION_STACK_INACTIVE and CLOUDFORMATION_STACK_NOT_EXISTING.

ResourceFilter

data ResourceFilter

A filter name and value pair that is used to obtain more specific results from a list of resources.

See: newResourceFilter smart constructor.

Constructors

ResourceFilter' ResourceFilterName (NonEmpty Text)

newResourceFilter

Arguments

:: ResourceFilterName
$sel:name:ResourceFilter'

-> NonEmpty Text
$sel:values:ResourceFilter'

-> ResourceFilter

Create a value of ResourceFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ResourceFilter', resourceFilter_name - The name of the filter. Filter names are case-sensitive.

$sel:values:ResourceFilter', resourceFilter_values - One or more filter values. Allowed filter values vary by resource filter name, and are case-sensitive.

ResourceIdentifier

data ResourceIdentifier

A structure that contains the ARN of a resource and its resource type.

See: newResourceIdentifier smart constructor.

Constructors

ResourceIdentifier' (Maybe Text) (Maybe Text)

newResourceIdentifier :: ResourceIdentifier

Create a value of ResourceIdentifier with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ResourceIdentifier', resourceIdentifier_resourceArn - The ARN of a resource.

$sel:resourceType:ResourceIdentifier', resourceIdentifier_resourceType - The resource type of a resource, such as AWS::EC2::Instance.

ResourceQuery

data ResourceQuery

The query that is used to define a resource group or a search for resources. A query specifies both a query type and a query string as a JSON object. See the examples section for example JSON strings.

The examples that follow are shown as standard JSON strings. If you include such a string as a parameter to the AWS CLI or an SDK API, you might need to 'escape' the string into a single line. For example, see the Quoting strings in the AWS CLI User Guide.

Example 1

The following generic example shows a resource query JSON string that includes only resources that meet the following criteria:

The resource type must be either resource_type1 or resource_type2.

The resource must have a tag Key1 with a value of either ValueA or ValueB.

The resource must have a tag Key2 with a value of either ValueC or ValueD.

{ "Type": "TAG_FILTERS_1_0", "Query": { "ResourceTypeFilters": [ "resource_type1", "resource_type2"], "TagFilters": [ { "Key": "Key1", "Values": ["ValueA","ValueB"] }, { "Key":"Key2", "Values":["ValueC","ValueD"] } ] } }

This has the equivalent "shortcut" syntax of the following:

{ "Type": "TAG_FILTERS_1_0", "Query": { "ResourceTypeFilters": [ "resource_type1", "resource_type2"], "TagFilters": [ { "Key1": ["ValueA","ValueB"] }, { "Key2": ["ValueC","ValueD"] } ] } }

Example 2

The following example shows a resource query JSON string that includes only Amazon EC2 instances that are tagged Stage with a value of Test.

{ "Type": "TAG_FILTERS_1_0", "Query": "{ "ResourceTypeFilters": "AWS::EC2::Instance", "TagFilters": { "Stage": "Test" } } }

Example 3

The following example shows a resource query JSON string that includes resource of any supported type as long as it is tagged Stage with a value of Prod.

{ "Type": "TAG_FILTERS_1_0", "Query": { "ResourceTypeFilters": "AWS::AllSupported", "TagFilters": { "Stage": "Prod" } } }

Example 4

The following example shows a resource query JSON string that includes only Amazon EC2 instances and Amazon S3 buckets that are part of the specified AWS CloudFormation stack.

{ "Type": "CLOUDFORMATION_STACK_1_0", "Query": { "ResourceTypeFilters": [ "AWS::EC2::Instance", "AWS::S3::Bucket" ], "StackIdentifier": "arn:aws:cloudformation:us-west-2:123456789012:stack/AWStestuseraccount/fb0d5000-aba8-00e8-aa9e-50d5cEXAMPLE" } }

See: newResourceQuery smart constructor.

Constructors

ResourceQuery' QueryType Text

newResourceQuery

Arguments

:: QueryType
$sel:type':ResourceQuery'

-> Text
$sel:query:ResourceQuery'

-> ResourceQuery

Create a value of ResourceQuery with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':ResourceQuery', resourceQuery_type - The type of the query. You can use the following values:

CLOUDFORMATION_STACK_1_0: Specifies that the Query contains an ARN for a CloudFormation stack.

TAG_FILTERS_1_0: Specifies that the Query parameter contains a JSON string that represents a collection of simple tag filters for resource types and tags. The JSON string uses a syntax similar to the GetResources operation, but uses only the ResourceTypeFilters and TagFilters fields. If you specify more than one tag key, only resources that match all tag keys, and at least one value of each specified tag key, are returned in your query. If you specify more than one value for a tag key, a resource matches the filter if it has a tag key value that matches any of the specified values.

For example, consider the following sample query for resources that have two tags, Stage and Version, with two values each:

[{"Stage":["Test","Deploy"]},{"Version":["1","2"]}]

The results of this query could include the following.

An EC2 instance that has the following two tags: {"Stage":"Deploy"}, and {"Version":"2"}

An S3 bucket that has the following two tags: {"Stage":"Test"}, and {"Version":"1"}

The query would not include the following items in the results, however.

An EC2 instance that has only the following tag: {"Stage":"Deploy"}.

The instance does not have all of the tag keys specified in the filter, so it is excluded from the results.

An RDS database that has the following two tags: {"Stage":"Archived"} and {"Version":"4"}

The database has all of the tag keys, but none of those keys has an associated value that matches at least one of the specified values in the filter.

$sel:query:ResourceQuery', resourceQuery_searchQuery - The query that defines a group or a search.

ResourceStatus

data ResourceStatus

A structure that identifies the current group membership status for a resource. Adding a resource to a resource group is performed asynchronously as a background task. A PENDING status indicates, for this resource, that the process isn't completed yet.

See: newResourceStatus smart constructor.

Constructors

ResourceStatus' (Maybe ResourceStatusValue)

newResourceStatus :: ResourceStatus

Create a value of ResourceStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ResourceStatus', resourceStatus_name - The current status.