Network.AWS.ResourceGroups.CreateGroup

Description

Creates a resource group with the specified name and description. You can optionally include a resource query, or a service configuration. For more information about constructing a resource query, see Create a tag-based group in Resource Groups. For more information about service configurations, see Service configurations for resource groups.

Minimum permissions

To run this command, you must have the following permissions:

resource-groups:CreateGroup

Creating a Request

data CreateGroup

See: newCreateGroup smart constructor.

Constructors

CreateGroup'

Fields

configuration :: Maybe [GroupConfigurationItem]

A configuration associates the resource group with an AWS service and specifies how the service can interact with the resources in the group. A configuration is an array of GroupConfigurationItem elements. For details about the syntax of service configurations, see Service configurations for resource groups.

A resource group can contain either a Configuration or a ResourceQuery, but not both.

tags :: Maybe (HashMap Text Text)

The tags to add to the group. A tag is key-value pair string.

description :: Maybe Text

The description of the resource group. Descriptions can consist of letters, numbers, hyphens, underscores, periods, and spaces.

resourceQuery :: Maybe ResourceQuery

The resource query that determines which AWS resources are members of this group. For more information about resource queries, see Create a tag-based group in Resource Groups.

A resource group can contain either a ResourceQuery or a Configuration, but not both.

name :: Text

The name of the group, which is the identifier of the group in other operations. You can't change the name of a resource group after you create it. A resource group name can consist of letters, numbers, hyphens, periods, and underscores. The name cannot start with AWS or aws; these are reserved. A resource group name must be unique within each AWS Region in your AWS account.

newCreateGroup

Arguments

:: Text
$sel:name:CreateGroup'

-> CreateGroup

Create a value of CreateGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configuration:CreateGroup', createGroup_configuration - A configuration associates the resource group with an AWS service and specifies how the service can interact with the resources in the group. A configuration is an array of GroupConfigurationItem elements. For details about the syntax of service configurations, see Service configurations for resource groups.

A resource group can contain either a Configuration or a ResourceQuery, but not both.

$sel:tags:CreateGroup', createGroup_tags - The tags to add to the group. A tag is key-value pair string.

$sel:description:CreateGroup', createGroup_description - The description of the resource group. Descriptions can consist of letters, numbers, hyphens, underscores, periods, and spaces.

$sel:resourceQuery:CreateGroup', createGroup_resourceQuery - The resource query that determines which AWS resources are members of this group. For more information about resource queries, see Create a tag-based group in Resource Groups.

A resource group can contain either a ResourceQuery or a Configuration, but not both.

$sel:name:CreateGroup', createGroup_name - The name of the group, which is the identifier of the group in other operations. You can't change the name of a resource group after you create it. A resource group name can consist of letters, numbers, hyphens, periods, and underscores. The name cannot start with AWS or aws; these are reserved. A resource group name must be unique within each AWS Region in your AWS account.

Request Lenses

createGroup_configuration :: Lens' CreateGroup (Maybe [GroupConfigurationItem])

A configuration associates the resource group with an AWS service and specifies how the service can interact with the resources in the group. A configuration is an array of GroupConfigurationItem elements. For details about the syntax of service configurations, see Service configurations for resource groups.

A resource group can contain either a Configuration or a ResourceQuery, but not both.

createGroup_tags :: Lens' CreateGroup (Maybe (HashMap Text Text))

The tags to add to the group. A tag is key-value pair string.

createGroup_description :: Lens' CreateGroup (Maybe Text)

The description of the resource group. Descriptions can consist of letters, numbers, hyphens, underscores, periods, and spaces.

createGroup_resourceQuery :: Lens' CreateGroup (Maybe ResourceQuery)

The resource query that determines which AWS resources are members of this group. For more information about resource queries, see Create a tag-based group in Resource Groups.

A resource group can contain either a ResourceQuery or a Configuration, but not both.

createGroup_name :: Lens' CreateGroup Text

The name of the group, which is the identifier of the group in other operations. You can't change the name of a resource group after you create it. A resource group name can consist of letters, numbers, hyphens, periods, and underscores. The name cannot start with AWS or aws; these are reserved. A resource group name must be unique within each AWS Region in your AWS account.

Destructuring the Response

data CreateGroupResponse

See: newCreateGroupResponse smart constructor.

Constructors

CreateGroupResponse'

Fields

groupConfiguration :: Maybe GroupConfiguration

The service configuration associated with the resource group. For details about the syntax of a service configuration, see Service configurations for resource groups.

group' :: Maybe Group

The description of the resource group.

tags :: Maybe (HashMap Text Text)

The tags associated with the group.

resourceQuery :: Maybe ResourceQuery

The resource query associated with the group. For more information about resource queries, see Create a tag-based group in Resource Groups.

httpStatus :: Int

The response's http status code.

newCreateGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateGroupResponse'

-> CreateGroupResponse

Create a value of CreateGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupConfiguration:CreateGroupResponse', createGroupResponse_groupConfiguration - The service configuration associated with the resource group. For details about the syntax of a service configuration, see Service configurations for resource groups.

$sel:group':CreateGroupResponse', createGroupResponse_group - The description of the resource group.

$sel:tags:CreateGroup', createGroupResponse_tags - The tags associated with the group.

$sel:resourceQuery:CreateGroup', createGroupResponse_resourceQuery - The resource query associated with the group. For more information about resource queries, see Create a tag-based group in Resource Groups.

$sel:httpStatus:CreateGroupResponse', createGroupResponse_httpStatus - The response's http status code.

Response Lenses

createGroupResponse_groupConfiguration :: Lens' CreateGroupResponse (Maybe GroupConfiguration)

The service configuration associated with the resource group. For details about the syntax of a service configuration, see Service configurations for resource groups.

createGroupResponse_group :: Lens' CreateGroupResponse (Maybe Group)

The description of the resource group.

createGroupResponse_tags :: Lens' CreateGroupResponse (Maybe (HashMap Text Text))

The tags associated with the group.

createGroupResponse_resourceQuery :: Lens' CreateGroupResponse (Maybe ResourceQuery)

The resource query associated with the group. For more information about resource queries, see Create a tag-based group in Resource Groups.

createGroupResponse_httpStatus :: Lens' CreateGroupResponse Int

The response's http status code.