Network.AWS.ResourceGroups.GetGroupConfiguration

Description

Returns the service configuration associated with the specified resource group. For details about the service configuration syntax, see Service configurations for resource groups.

Minimum permissions

To run this command, you must have the following permissions:

resource-groups:GetGroupConfiguration

Creating a Request

data GetGroupConfiguration

See: newGetGroupConfiguration smart constructor.

Constructors

GetGroupConfiguration'

Fields

group' :: Maybe Text

The name or the ARN of the resource group.

newGetGroupConfiguration :: GetGroupConfiguration

Create a value of GetGroupConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:group':GetGroupConfiguration', getGroupConfiguration_group - The name or the ARN of the resource group.

Request Lenses

getGroupConfiguration_group :: Lens' GetGroupConfiguration (Maybe Text)

The name or the ARN of the resource group.

Destructuring the Response

data GetGroupConfigurationResponse

See: newGetGroupConfigurationResponse smart constructor.

Constructors

GetGroupConfigurationResponse'

Fields

groupConfiguration :: Maybe GroupConfiguration

The service configuration associated with the specified group. For details about the service configuration syntax, see Service configurations for resource groups.

httpStatus :: Int

The response's http status code.

newGetGroupConfigurationResponse

Arguments

:: Int
$sel:httpStatus:GetGroupConfigurationResponse'

-> GetGroupConfigurationResponse

Create a value of GetGroupConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupConfiguration:GetGroupConfigurationResponse', getGroupConfigurationResponse_groupConfiguration - The service configuration associated with the specified group. For details about the service configuration syntax, see Service configurations for resource groups.

$sel:httpStatus:GetGroupConfigurationResponse', getGroupConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

getGroupConfigurationResponse_groupConfiguration :: Lens' GetGroupConfigurationResponse (Maybe GroupConfiguration)

The service configuration associated with the specified group. For details about the service configuration syntax, see Service configurations for resource groups.

getGroupConfigurationResponse_httpStatus :: Lens' GetGroupConfigurationResponse Int

The response's http status code.