Network.AWS.ResourceGroups.GetGroupQuery

Description

Retrieves the resource query associated with the specified resource group. For more information about resource queries, see Create a tag-based group in Resource Groups.

Minimum permissions

To run this command, you must have the following permissions:

resource-groups:GetGroupQuery

Creating a Request

data GetGroupQuery

See: newGetGroupQuery smart constructor.

Constructors

GetGroupQuery'

Fields

group' :: Maybe Text

The name or the ARN of the resource group to query.

groupName :: Maybe Text

Don't use this parameter. Use Group instead.

newGetGroupQuery :: GetGroupQuery

Create a value of GetGroupQuery with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:group':GetGroupQuery', getGroupQuery_group - The name or the ARN of the resource group to query.

$sel:groupName:GetGroupQuery', getGroupQuery_groupName - Don't use this parameter. Use Group instead.

Request Lenses

getGroupQuery_group :: Lens' GetGroupQuery (Maybe Text)

The name or the ARN of the resource group to query.

getGroupQuery_groupName :: Lens' GetGroupQuery (Maybe Text)

Don't use this parameter. Use Group instead.

Destructuring the Response

data GetGroupQueryResponse

See: newGetGroupQueryResponse smart constructor.

Constructors

GetGroupQueryResponse'

Fields

groupQuery :: Maybe GroupQuery

The resource query associated with the specified group. For more information about resource queries, see Create a tag-based group in Resource Groups.

httpStatus :: Int

The response's http status code.

newGetGroupQueryResponse

Arguments

:: Int
$sel:httpStatus:GetGroupQueryResponse'

-> GetGroupQueryResponse

Create a value of GetGroupQueryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupQuery:GetGroupQueryResponse', getGroupQueryResponse_groupQuery - The resource query associated with the specified group. For more information about resource queries, see Create a tag-based group in Resource Groups.

$sel:httpStatus:GetGroupQueryResponse', getGroupQueryResponse_httpStatus - The response's http status code.

Response Lenses

getGroupQueryResponse_groupQuery :: Lens' GetGroupQueryResponse (Maybe GroupQuery)

The resource query associated with the specified group. For more information about resource queries, see Create a tag-based group in Resource Groups.

getGroupQueryResponse_httpStatus :: Lens' GetGroupQueryResponse Int

The response's http status code.