Network.AWS.ResourceGroups.Lens

Description

Operations

GetGroupConfiguration

getGroupConfiguration_group :: Lens' GetGroupConfiguration (Maybe Text)

The name or the ARN of the resource group.

getGroupConfigurationResponse_groupConfiguration :: Lens' GetGroupConfigurationResponse (Maybe GroupConfiguration)

The service configuration associated with the specified group. For details about the service configuration syntax, see Service configurations for resource groups.

getGroupConfigurationResponse_httpStatus :: Lens' GetGroupConfigurationResponse Int

The response's http status code.

PutGroupConfiguration

putGroupConfiguration_configuration :: Lens' PutGroupConfiguration (Maybe [GroupConfigurationItem])

The new configuration to associate with the specified group. A configuration associates the resource group with an AWS service and specifies how the service can interact with the resources in the group. A configuration is an array of GroupConfigurationItem elements.

For information about the syntax of a service configuration, see Service configurations for resource groups.

A resource group can contain either a Configuration or a ResourceQuery, but not both.

putGroupConfiguration_group :: Lens' PutGroupConfiguration (Maybe Text)

The name or ARN of the resource group with the configuration that you want to update.

putGroupConfigurationResponse_httpStatus :: Lens' PutGroupConfigurationResponse Int

The response's http status code.

ListGroups

listGroups_nextToken :: Lens' ListGroups (Maybe Text)

The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value provided by a previous call's NextToken response to indicate where the output should continue from.

listGroups_maxResults :: Lens' ListGroups (Maybe Natural)

The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that the service might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

listGroups_filters :: Lens' ListGroups (Maybe [GroupFilter])

Filters, formatted as GroupFilter objects, that you want to apply to a ListGroups operation.

resource-type - Filter the results to include only those of the specified resource types. Specify up to five resource types in the format AWS::ServiceCode::ResourceType . For example, AWS::EC2::Instance, or AWS::S3::Bucket.

configuration-type - Filter the results to include only those groups that have the specified configuration types attached. The current supported values are:

AWS:EC2::CapacityReservationPool

AWS:EC2::HostManagement

listGroupsResponse_groups :: Lens' ListGroupsResponse (Maybe [Group])

/__Deprecated - don't use this field. Use the GroupIdentifiers response field instead.__/

listGroupsResponse_nextToken :: Lens' ListGroupsResponse (Maybe Text)

If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

listGroupsResponse_groupIdentifiers :: Lens' ListGroupsResponse (Maybe [GroupIdentifier])

A list of GroupIdentifier objects. Each identifier is an object that contains both the Name and the GroupArn.

listGroupsResponse_httpStatus :: Lens' ListGroupsResponse Int

The response's http status code.

CreateGroup

createGroup_configuration :: Lens' CreateGroup (Maybe [GroupConfigurationItem])

A configuration associates the resource group with an AWS service and specifies how the service can interact with the resources in the group. A configuration is an array of GroupConfigurationItem elements. For details about the syntax of service configurations, see Service configurations for resource groups.

A resource group can contain either a Configuration or a ResourceQuery, but not both.

createGroup_tags :: Lens' CreateGroup (Maybe (HashMap Text Text))

The tags to add to the group. A tag is key-value pair string.

createGroup_description :: Lens' CreateGroup (Maybe Text)

The description of the resource group. Descriptions can consist of letters, numbers, hyphens, underscores, periods, and spaces.

createGroup_resourceQuery :: Lens' CreateGroup (Maybe ResourceQuery)

The resource query that determines which AWS resources are members of this group. For more information about resource queries, see Create a tag-based group in Resource Groups.

A resource group can contain either a ResourceQuery or a Configuration, but not both.

createGroup_name :: Lens' CreateGroup Text

The name of the group, which is the identifier of the group in other operations. You can't change the name of a resource group after you create it. A resource group name can consist of letters, numbers, hyphens, periods, and underscores. The name cannot start with AWS or aws; these are reserved. A resource group name must be unique within each AWS Region in your AWS account.

createGroupResponse_groupConfiguration :: Lens' CreateGroupResponse (Maybe GroupConfiguration)

The service configuration associated with the resource group. For details about the syntax of a service configuration, see Service configurations for resource groups.

createGroupResponse_group :: Lens' CreateGroupResponse (Maybe Group)

The description of the resource group.

createGroupResponse_tags :: Lens' CreateGroupResponse (Maybe (HashMap Text Text))

The tags associated with the group.

createGroupResponse_resourceQuery :: Lens' CreateGroupResponse (Maybe ResourceQuery)

The resource query associated with the group. For more information about resource queries, see Create a tag-based group in Resource Groups.

createGroupResponse_httpStatus :: Lens' CreateGroupResponse Int

The response's http status code.

GetGroupQuery

getGroupQuery_groupName :: Lens' GetGroupQuery (Maybe Text)

Don't use this parameter. Use Group instead.

getGroupQuery_group :: Lens' GetGroupQuery (Maybe Text)

The name or the ARN of the resource group to query.

getGroupQueryResponse_groupQuery :: Lens' GetGroupQueryResponse (Maybe GroupQuery)

The resource query associated with the specified group. For more information about resource queries, see Create a tag-based group in Resource Groups.

getGroupQueryResponse_httpStatus :: Lens' GetGroupQueryResponse Int

The response's http status code.

SearchResources

searchResources_nextToken :: Lens' SearchResources (Maybe Text)

The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value provided by a previous call's NextToken response to indicate where the output should continue from.

searchResources_maxResults :: Lens' SearchResources (Maybe Natural)

The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that the service might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

searchResources_resourceQuery :: Lens' SearchResources ResourceQuery

The search query, using the same formats that are supported for resource group definition. For more information, see CreateGroup.

searchResourcesResponse_nextToken :: Lens' SearchResourcesResponse (Maybe Text)

If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

searchResourcesResponse_resourceIdentifiers :: Lens' SearchResourcesResponse (Maybe [ResourceIdentifier])

The ARNs and resource types of resources that are members of the group that you specified.

searchResourcesResponse_queryErrors :: Lens' SearchResourcesResponse (Maybe [QueryError])

A list of QueryError objects. Each error is an object that contains ErrorCode and Message structures. Possible values for ErrorCode are CLOUDFORMATION_STACK_INACTIVE and CLOUDFORMATION_STACK_NOT_EXISTING.

searchResourcesResponse_httpStatus :: Lens' SearchResourcesResponse Int

The response's http status code.

GetTags

getTags_arn :: Lens' GetTags Text

The ARN of the resource group whose tags you want to retrieve.

getTagsResponse_arn :: Lens' GetTagsResponse (Maybe Text)

The ARN of the tagged resource group.

getTagsResponse_tags :: Lens' GetTagsResponse (Maybe (HashMap Text Text))

The tags associated with the specified resource group.

getTagsResponse_httpStatus :: Lens' GetTagsResponse Int

The response's http status code.

UpdateGroupQuery

updateGroupQuery_groupName :: Lens' UpdateGroupQuery (Maybe Text)

Don't use this parameter. Use Group instead.

updateGroupQuery_group :: Lens' UpdateGroupQuery (Maybe Text)

The name or the ARN of the resource group to query.

updateGroupQuery_resourceQuery :: Lens' UpdateGroupQuery ResourceQuery

The resource query to determine which AWS resources are members of this resource group.

A resource group can contain either a Configuration or a ResourceQuery, but not both.

updateGroupQueryResponse_groupQuery :: Lens' UpdateGroupQueryResponse (Maybe GroupQuery)

The updated resource query associated with the resource group after the update.

updateGroupQueryResponse_httpStatus :: Lens' UpdateGroupQueryResponse Int

The response's http status code.

ListGroupResources

listGroupResources_nextToken :: Lens' ListGroupResources (Maybe Text)

The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value provided by a previous call's NextToken response to indicate where the output should continue from.

listGroupResources_maxResults :: Lens' ListGroupResources (Maybe Natural)

The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that the service might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

listGroupResources_groupName :: Lens' ListGroupResources (Maybe Text)

/__Deprecated - don't use this parameter. Use the Group request field instead.__/

listGroupResources_group :: Lens' ListGroupResources (Maybe Text)

The name or the ARN of the resource group

listGroupResources_filters :: Lens' ListGroupResources (Maybe [ResourceFilter])

Filters, formatted as ResourceFilter objects, that you want to apply to a ListGroupResources operation. Filters the results to include only those of the specified resource types.

resource-type - Filter resources by their type. Specify up to five resource types in the format AWS::ServiceCode::ResourceType. For example, AWS::EC2::Instance, or AWS::S3::Bucket.

When you specify a resource-type filter for ListGroupResources, AWS Resource Groups validates your filter resource types against the types that are defined in the query associated with the group. For example, if a group contains only S3 buckets because its query specifies only that resource type, but your resource-type filter includes EC2 instances, AWS Resource Groups does not filter for EC2 instances. In this case, a ListGroupResources request returns a BadRequestException error with a message similar to the following:

The resource types specified as filters in the request are not valid.

The error includes a list of resource types that failed the validation because they are not part of the query associated with the group. This validation doesn't occur when the group query specifies AWS::AllSupported, because a group based on such a query can contain any of the allowed resource types for the query type (tag-based or AWS CloudFormation stack-based queries).

listGroupResourcesResponse_nextToken :: Lens' ListGroupResourcesResponse (Maybe Text)

If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

listGroupResourcesResponse_resourceIdentifiers :: Lens' ListGroupResourcesResponse (Maybe [ResourceIdentifier])

__/Deprecated - don't use this parameter. Use the Resources response field instead./__

listGroupResourcesResponse_queryErrors :: Lens' ListGroupResourcesResponse (Maybe [QueryError])

A list of QueryError objects. Each error is an object that contains ErrorCode and Message structures. Possible values for ErrorCode are CLOUDFORMATION_STACK_INACTIVE and CLOUDFORMATION_STACK_NOT_EXISTING.

listGroupResourcesResponse_resources :: Lens' ListGroupResourcesResponse (Maybe [ListGroupResourcesItem])

An array of resources from which you can determine each resource's identity, type, and group membership status.

listGroupResourcesResponse_httpStatus :: Lens' ListGroupResourcesResponse Int

The response's http status code.

Untag

untag_arn :: Lens' Untag Text

The ARN of the resource group from which to remove tags. The command removed both the specified keys and any values associated with those keys.

untag_keys :: Lens' Untag [Text]

The keys of the tags to be removed.

untagResponse_arn :: Lens' UntagResponse (Maybe Text)

The ARN of the resource group from which tags have been removed.

untagResponse_keys :: Lens' UntagResponse (Maybe [Text])

The keys of the tags that were removed.

untagResponse_httpStatus :: Lens' UntagResponse Int

The response's http status code.

GetGroup

getGroup_groupName :: Lens' GetGroup (Maybe Text)

Deprecated - don't use this parameter. Use Group instead.

getGroup_group :: Lens' GetGroup (Maybe Text)

The name or the ARN of the resource group to retrieve.

getGroupResponse_group :: Lens' GetGroupResponse (Maybe Group)

A full description of the resource group.

getGroupResponse_httpStatus :: Lens' GetGroupResponse Int

The response's http status code.

DeleteGroup

deleteGroup_groupName :: Lens' DeleteGroup (Maybe Text)

Deprecated - don't use this parameter. Use Group instead.

deleteGroup_group :: Lens' DeleteGroup (Maybe Text)

The name or the ARN of the resource group to delete.

deleteGroupResponse_group :: Lens' DeleteGroupResponse (Maybe Group)

A full description of the deleted resource group.

deleteGroupResponse_httpStatus :: Lens' DeleteGroupResponse Int

The response's http status code.

GroupResources

groupResources_group :: Lens' GroupResources Text

The name or the ARN of the resource group to add resources to.

groupResources_resourceArns :: Lens' GroupResources (NonEmpty Text)

The list of ARNs for resources to be added to the group.

groupResourcesResponse_succeeded :: Lens' GroupResourcesResponse (Maybe (NonEmpty Text))

A list of ARNs of resources that were successfully added to the group by this operation.

groupResourcesResponse_pending :: Lens' GroupResourcesResponse (Maybe [PendingResource])

A list of ARNs of any resources that are still in the process of being added to the group by this operation. These pending additions continue asynchronously. You can check the status of pending additions by using the ListGroupResources operation, and checking the Resources array in the response and the Status field of each object in that array.

groupResourcesResponse_failed :: Lens' GroupResourcesResponse (Maybe [FailedResource])

A list of ARNs of any resources that failed to be added to the group by this operation.

groupResourcesResponse_httpStatus :: Lens' GroupResourcesResponse Int

The response's http status code.

UpdateGroup

updateGroup_groupName :: Lens' UpdateGroup (Maybe Text)

Don't use this parameter. Use Group instead.

updateGroup_group :: Lens' UpdateGroup (Maybe Text)

The name or the ARN of the resource group to modify.

updateGroup_description :: Lens' UpdateGroup (Maybe Text)

The new description that you want to update the resource group with. Descriptions can contain letters, numbers, hyphens, underscores, periods, and spaces.

updateGroupResponse_group :: Lens' UpdateGroupResponse (Maybe Group)

The update description of the resource group.

updateGroupResponse_httpStatus :: Lens' UpdateGroupResponse Int

The response's http status code.

UngroupResources

ungroupResources_group :: Lens' UngroupResources Text

The name or the ARN of the resource group from which to remove the resources.

ungroupResources_resourceArns :: Lens' UngroupResources (NonEmpty Text)

The ARNs of the resources to be removed from the group.

ungroupResourcesResponse_succeeded :: Lens' UngroupResourcesResponse (Maybe (NonEmpty Text))

A list of resources that were successfully removed from the group by this operation.

ungroupResourcesResponse_pending :: Lens' UngroupResourcesResponse (Maybe [PendingResource])

A list of any resources that are still in the process of being removed from the group by this operation. These pending removals continue asynchronously. You can check the status of pending removals by using the ListGroupResources operation. After the resource is successfully removed, it no longer appears in the response.

ungroupResourcesResponse_failed :: Lens' UngroupResourcesResponse (Maybe [FailedResource])

A list of any resources that failed to be removed from the group by this operation.

ungroupResourcesResponse_httpStatus :: Lens' UngroupResourcesResponse Int

The response's http status code.

Tag

tag_arn :: Lens' Tag Text

The ARN of the resource group to which to add tags.

tag_tags :: Lens' Tag (HashMap Text Text)

The tags to add to the specified resource group. A tag is a string-to-string map of key-value pairs.

tagResponse_arn :: Lens' TagResponse (Maybe Text)

The ARN of the tagged resource.

tagResponse_tags :: Lens' TagResponse (Maybe (HashMap Text Text))

The tags that have been added to the specified resource group.

tagResponse_httpStatus :: Lens' TagResponse Int

The response's http status code.

Types

FailedResource

failedResource_resourceArn :: Lens' FailedResource (Maybe Text)

The ARN of the resource that failed to be added or removed.

failedResource_errorMessage :: Lens' FailedResource (Maybe Text)

The error message text associated with the failure.

failedResource_errorCode :: Lens' FailedResource (Maybe Text)

The error code associated with the failure.

Group

group_description :: Lens' Group (Maybe Text)

The description of the resource group.

group_groupArn :: Lens' Group Text

The ARN of the resource group.

group_name :: Lens' Group Text

The name of the resource group.

GroupConfiguration

groupConfiguration_status :: Lens' GroupConfiguration (Maybe GroupConfigurationStatus)

The current status of an attempt to update the group configuration.

groupConfiguration_configuration :: Lens' GroupConfiguration (Maybe [GroupConfigurationItem])

The configuration currently associated with the group and in effect.

groupConfiguration_failureReason :: Lens' GroupConfiguration (Maybe Text)

If present, the reason why a request to update the group configuration failed.

groupConfiguration_proposedConfiguration :: Lens' GroupConfiguration (Maybe [GroupConfigurationItem])

If present, the new configuration that is in the process of being applied to the group.

GroupConfigurationItem

groupConfigurationItem_parameters :: Lens' GroupConfigurationItem (Maybe [GroupConfigurationParameter])

A collection of parameters for this group configuration item. For the list of parameters that you can use with each configuration item type, see Supported resource types and parameters.

groupConfigurationItem_type :: Lens' GroupConfigurationItem Text

Specifies the type of group configuration item. Each item must have a unique value for type. For the list of types that you can specify for a configuration item, see Supported resource types and parameters.

GroupConfigurationParameter

groupConfigurationParameter_values :: Lens' GroupConfigurationParameter (Maybe [Text])

The value or values to be used for the specified parameter. For the list of values you can use with each parameter, see Supported resource types and parameters.

groupConfigurationParameter_name :: Lens' GroupConfigurationParameter Text

The name of the group configuration parameter. For the list of parameters that you can use with each configuration item type, see Supported resource types and parameters.

GroupFilter

groupFilter_name :: Lens' GroupFilter GroupFilterName

The name of the filter. Filter names are case-sensitive.

groupFilter_values :: Lens' GroupFilter (NonEmpty Text)

One or more filter values. Allowed filter values vary by group filter name, and are case-sensitive.

GroupIdentifier

groupIdentifier_groupName :: Lens' GroupIdentifier (Maybe Text)

The name of the resource group.

groupIdentifier_groupArn :: Lens' GroupIdentifier (Maybe Text)

The ARN of the resource group.

GroupQuery

groupQuery_groupName :: Lens' GroupQuery Text

The name of the resource group that is associated with the specified resource query.

groupQuery_resourceQuery :: Lens' GroupQuery ResourceQuery

The resource query that determines which AWS resources are members of the associated resource group.

ListGroupResourcesItem

listGroupResourcesItem_status :: Lens' ListGroupResourcesItem (Maybe ResourceStatus)

A structure that contains the status of this resource's membership in the group.

This field is present in the response only if the group is of type AWS::EC2::HostManagement.

listGroupResourcesItem_identifier :: Lens' ListGroupResourcesItem (Maybe ResourceIdentifier)

Undocumented member.

PendingResource

pendingResource_resourceArn :: Lens' PendingResource (Maybe Text)

The Amazon resource name (ARN) of the resource that's in a pending state.

QueryError

queryError_message :: Lens' QueryError (Maybe Text)

A message that explains the ErrorCode value. Messages might state that the specified CloudFormation stack does not exist (or no longer exists). For CLOUDFORMATION_STACK_INACTIVE, the message typically states that the CloudFormation stack has a status that is not (or no longer) active, such as CREATE_FAILED.

queryError_errorCode :: Lens' QueryError (Maybe QueryErrorCode)

Possible values are CLOUDFORMATION_STACK_INACTIVE and CLOUDFORMATION_STACK_NOT_EXISTING.

ResourceFilter

resourceFilter_name :: Lens' ResourceFilter ResourceFilterName

The name of the filter. Filter names are case-sensitive.

resourceFilter_values :: Lens' ResourceFilter (NonEmpty Text)

One or more filter values. Allowed filter values vary by resource filter name, and are case-sensitive.

ResourceIdentifier

resourceIdentifier_resourceArn :: Lens' ResourceIdentifier (Maybe Text)

The ARN of a resource.

resourceIdentifier_resourceType :: Lens' ResourceIdentifier (Maybe Text)

The resource type of a resource, such as AWS::EC2::Instance.

ResourceQuery

resourceQuery_type :: Lens' ResourceQuery QueryType

The type of the query. You can use the following values:

CLOUDFORMATION_STACK_1_0: Specifies that the Query contains an ARN for a CloudFormation stack.

TAG_FILTERS_1_0: Specifies that the Query parameter contains a JSON string that represents a collection of simple tag filters for resource types and tags. The JSON string uses a syntax similar to the GetResources operation, but uses only the ResourceTypeFilters and TagFilters fields. If you specify more than one tag key, only resources that match all tag keys, and at least one value of each specified tag key, are returned in your query. If you specify more than one value for a tag key, a resource matches the filter if it has a tag key value that matches any of the specified values.

For example, consider the following sample query for resources that have two tags, Stage and Version, with two values each:

[{"Stage":["Test","Deploy"]},{"Version":["1","2"]}]

The results of this query could include the following.

An EC2 instance that has the following two tags: {"Stage":"Deploy"}, and {"Version":"2"}

An S3 bucket that has the following two tags: {"Stage":"Test"}, and {"Version":"1"}

The query would not include the following items in the results, however.

An EC2 instance that has only the following tag: {"Stage":"Deploy"}.

The instance does not have all of the tag keys specified in the filter, so it is excluded from the results.

An RDS database that has the following two tags: {"Stage":"Archived"} and {"Version":"4"}

The database has all of the tag keys, but none of those keys has an associated value that matches at least one of the specified values in the filter.

resourceQuery_searchQuery :: Lens' ResourceQuery Text

The query that defines a group or a search.

ResourceStatus

resourceStatus_name :: Lens' ResourceStatus (Maybe ResourceStatusValue)

The current status.