Network.AWS.ResourceGroups.Tag

Description

Adds tags to a resource group with the specified ARN. Existing tags on a resource group are not changed if they are not specified in the request parameters.

Do not store personally identifiable information (PII) or other confidential or sensitive information in tags. We use tags to provide you with billing and administration services. Tags are not intended to be used for private or sensitive data.

Minimum permissions

To run this command, you must have the following permissions:

resource-groups:Tag

Creating a Request

data Tag

See: newTag smart constructor.

Constructors

Tag'

Fields

arn :: Text

The ARN of the resource group to which to add tags.

tags :: HashMap Text Text

The tags to add to the specified resource group. A tag is a string-to-string map of key-value pairs.

newTag

Arguments

:: Text
$sel:arn:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Tag', tag_arn - The ARN of the resource group to which to add tags.

$sel:tags:Tag', tag_tags - The tags to add to the specified resource group. A tag is a string-to-string map of key-value pairs.

Request Lenses

tag_arn :: Lens' Tag Text

The ARN of the resource group to which to add tags.

tag_tags :: Lens' Tag (HashMap Text Text)

The tags to add to the specified resource group. A tag is a string-to-string map of key-value pairs.

Destructuring the Response

data TagResponse

See: newTagResponse smart constructor.

Constructors

TagResponse'

Fields

arn :: Maybe Text

The ARN of the tagged resource.

tags :: Maybe (HashMap Text Text)

The tags that have been added to the specified resource group.

httpStatus :: Int

The response's http status code.

newTagResponse

Arguments

:: Int
$sel:httpStatus:TagResponse'

-> TagResponse

Create a value of TagResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Tag', tagResponse_arn - The ARN of the tagged resource.

$sel:tags:Tag', tagResponse_tags - The tags that have been added to the specified resource group.

$sel:httpStatus:TagResponse', tagResponse_httpStatus - The response's http status code.

Response Lenses

tagResponse_arn :: Lens' TagResponse (Maybe Text)

The ARN of the tagged resource.

tagResponse_tags :: Lens' TagResponse (Maybe (HashMap Text Text))

The tags that have been added to the specified resource group.

tagResponse_httpStatus :: Lens' TagResponse Int

The response's http status code.