Network.AWS.ResourceGroups.Types

Description

Service Configuration

defaultService :: Service

API version 2017-11-27 of the Amazon Resource Groups SDK configuration.

Errors

_ForbiddenException :: AsError a => Getting (First ServiceError) a ServiceError

The caller isn't authorized to make the request. Check permissions.

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

One or more of the specified resources don't exist.

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

You've exceeded throttling limits by making too many requests in a period of time.

_InternalServerErrorException :: AsError a => Getting (First ServiceError) a ServiceError

An internal error occurred while processing the request. Try again later.

_MethodNotAllowedException :: AsError a => Getting (First ServiceError) a ServiceError

The request uses an HTTP method that isn't allowed for the specified resource.

_UnauthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because it doesn't have valid credentials for the target resource.

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request includes one or more parameters that violate validation rules.

GroupConfigurationStatus

newtype GroupConfigurationStatus

Constructors

GroupConfigurationStatus'

Fields

fromGroupConfigurationStatus :: Text

Bundled Patterns

pattern GroupConfigurationStatus_UPDATE_COMPLETE :: GroupConfigurationStatus

pattern GroupConfigurationStatus_UPDATE_FAILED :: GroupConfigurationStatus

pattern GroupConfigurationStatus_UPDATING :: GroupConfigurationStatus

GroupFilterName

newtype GroupFilterName

Constructors

GroupFilterName'

Fields

fromGroupFilterName :: Text

Bundled Patterns

pattern GroupFilterName_Configuration_type :: GroupFilterName

pattern GroupFilterName_Resource_type :: GroupFilterName

QueryErrorCode

newtype QueryErrorCode

Constructors

QueryErrorCode'

Fields

fromQueryErrorCode :: Text

Bundled Patterns

pattern QueryErrorCode_CLOUDFORMATION_STACK_INACTIVE :: QueryErrorCode

pattern QueryErrorCode_CLOUDFORMATION_STACK_NOT_EXISTING :: QueryErrorCode

QueryType

newtype QueryType

Constructors

QueryType'

Fields

fromQueryType :: Text

Bundled Patterns

pattern QueryType_CLOUDFORMATION_STACK_1_0 :: QueryType

pattern QueryType_TAG_FILTERS_1_0 :: QueryType

ResourceFilterName

newtype ResourceFilterName

Constructors

ResourceFilterName'

Fields

fromResourceFilterName :: Text

Bundled Patterns

pattern ResourceFilterName_Resource_type :: ResourceFilterName

ResourceStatusValue

newtype ResourceStatusValue

Constructors

ResourceStatusValue'

Fields

fromResourceStatusValue :: Text

Bundled Patterns

pattern ResourceStatusValue_PENDING :: ResourceStatusValue

FailedResource

data FailedResource

A resource that failed to be added to or removed from a group.

See: newFailedResource smart constructor.

Constructors

FailedResource'

Fields

resourceArn :: Maybe Text

The ARN of the resource that failed to be added or removed.

errorCode :: Maybe Text

The error code associated with the failure.

errorMessage :: Maybe Text

The error message text associated with the failure.

newFailedResource :: FailedResource

Create a value of FailedResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:FailedResource', failedResource_resourceArn - The ARN of the resource that failed to be added or removed.

$sel:errorCode:FailedResource', failedResource_errorCode - The error code associated with the failure.

$sel:errorMessage:FailedResource', failedResource_errorMessage - The error message text associated with the failure.

failedResource_resourceArn :: Lens' FailedResource (Maybe Text)

The ARN of the resource that failed to be added or removed.

failedResource_errorCode :: Lens' FailedResource (Maybe Text)

The error code associated with the failure.

failedResource_errorMessage :: Lens' FailedResource (Maybe Text)

The error message text associated with the failure.

Group

data Group

A resource group that contains AWS resources. You can assign resources to the group by associating either of the following elements with the group:

ResourceQuery - Use a resource query to specify a set of tag keys and values. All resources in the same AWS Region and AWS account that have those keys with the same values are included in the group. You can add a resource query when you create the group, or later by using the PutGroupConfiguration operation.

GroupConfiguration - Use a service configuration to associate the group with an AWS service. The configuration specifies which resource types can be included in the group.

See: newGroup smart constructor.

Constructors

Group'

Fields

description :: Maybe Text

The description of the resource group.

groupArn :: Text

The ARN of the resource group.

name :: Text

The name of the resource group.

newGroup

Arguments

:: Text
$sel:groupArn:Group'

-> Text
$sel:name:Group'

-> Group

Create a value of Group with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:Group', group_description - The description of the resource group.

$sel:groupArn:Group', group_groupArn - The ARN of the resource group.

$sel:name:Group', group_name - The name of the resource group.

group_description :: Lens' Group (Maybe Text)

The description of the resource group.

group_groupArn :: Lens' Group Text

The ARN of the resource group.

group_name :: Lens' Group Text

The name of the resource group.

GroupConfiguration

data GroupConfiguration

A service configuration associated with a resource group. The configuration options are determined by the AWS service that defines the Type, and specifies which resources can be included in the group. You can add a service configuration when you create the group by using CreateGroup, or later by using the PutGroupConfiguration operation. For details about group service configuration syntax, see Service configurations for resource groups.

See: newGroupConfiguration smart constructor.

Constructors

GroupConfiguration'

Fields

status :: Maybe GroupConfigurationStatus

The current status of an attempt to update the group configuration.

failureReason :: Maybe Text

If present, the reason why a request to update the group configuration failed.

proposedConfiguration :: Maybe [GroupConfigurationItem]

If present, the new configuration that is in the process of being applied to the group.

configuration :: Maybe [GroupConfigurationItem]

The configuration currently associated with the group and in effect.

newGroupConfiguration :: GroupConfiguration

Create a value of GroupConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GroupConfiguration', groupConfiguration_status - The current status of an attempt to update the group configuration.

$sel:failureReason:GroupConfiguration', groupConfiguration_failureReason - If present, the reason why a request to update the group configuration failed.

$sel:proposedConfiguration:GroupConfiguration', groupConfiguration_proposedConfiguration - If present, the new configuration that is in the process of being applied to the group.

$sel:configuration:GroupConfiguration', groupConfiguration_configuration - The configuration currently associated with the group and in effect.

groupConfiguration_status :: Lens' GroupConfiguration (Maybe GroupConfigurationStatus)

The current status of an attempt to update the group configuration.

groupConfiguration_failureReason :: Lens' GroupConfiguration (Maybe Text)

If present, the reason why a request to update the group configuration failed.

groupConfiguration_proposedConfiguration :: Lens' GroupConfiguration (Maybe [GroupConfigurationItem])

If present, the new configuration that is in the process of being applied to the group.

groupConfiguration_configuration :: Lens' GroupConfiguration (Maybe [GroupConfigurationItem])

The configuration currently associated with the group and in effect.

GroupConfigurationItem

data GroupConfigurationItem

An item in a group configuration. A group service configuration can have one or more items. For details about group service configuration syntax, see Service configurations for resource groups.

See: newGroupConfigurationItem smart constructor.

Constructors

GroupConfigurationItem'

Fields

parameters :: Maybe [GroupConfigurationParameter]

A collection of parameters for this group configuration item. For the list of parameters that you can use with each configuration item type, see Supported resource types and parameters.

type' :: Text

Specifies the type of group configuration item. Each item must have a unique value for type. For the list of types that you can specify for a configuration item, see Supported resource types and parameters.

newGroupConfigurationItem

Arguments

:: Text
$sel:type':GroupConfigurationItem'

-> GroupConfigurationItem

Create a value of GroupConfigurationItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameters:GroupConfigurationItem', groupConfigurationItem_parameters - A collection of parameters for this group configuration item. For the list of parameters that you can use with each configuration item type, see Supported resource types and parameters.

$sel:type':GroupConfigurationItem', groupConfigurationItem_type - Specifies the type of group configuration item. Each item must have a unique value for type. For the list of types that you can specify for a configuration item, see Supported resource types and parameters.

groupConfigurationItem_parameters :: Lens' GroupConfigurationItem (Maybe [GroupConfigurationParameter])

A collection of parameters for this group configuration item. For the list of parameters that you can use with each configuration item type, see Supported resource types and parameters.

groupConfigurationItem_type :: Lens' GroupConfigurationItem Text

Specifies the type of group configuration item. Each item must have a unique value for type. For the list of types that you can specify for a configuration item, see Supported resource types and parameters.

GroupConfigurationParameter

data GroupConfigurationParameter

A parameter for a group configuration item. For details about group service configuration syntax, see Service configurations for resource groups.

See: newGroupConfigurationParameter smart constructor.

Constructors

GroupConfigurationParameter'

Fields

values :: Maybe [Text]

The value or values to be used for the specified parameter. For the list of values you can use with each parameter, see Supported resource types and parameters.

name :: Text

The name of the group configuration parameter. For the list of parameters that you can use with each configuration item type, see Supported resource types and parameters.

newGroupConfigurationParameter

Arguments

:: Text
$sel:name:GroupConfigurationParameter'

-> GroupConfigurationParameter

Create a value of GroupConfigurationParameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:GroupConfigurationParameter', groupConfigurationParameter_values - The value or values to be used for the specified parameter. For the list of values you can use with each parameter, see Supported resource types and parameters.

$sel:name:GroupConfigurationParameter', groupConfigurationParameter_name - The name of the group configuration parameter. For the list of parameters that you can use with each configuration item type, see Supported resource types and parameters.

groupConfigurationParameter_values :: Lens' GroupConfigurationParameter (Maybe [Text])

The value or values to be used for the specified parameter. For the list of values you can use with each parameter, see Supported resource types and parameters.

groupConfigurationParameter_name :: Lens' GroupConfigurationParameter Text

The name of the group configuration parameter. For the list of parameters that you can use with each configuration item type, see Supported resource types and parameters.

GroupFilter

data GroupFilter

A filter collection that you can use to restrict the results from a List operation to only those you want to include.

See: newGroupFilter smart constructor.

Constructors

GroupFilter'

Fields

name :: GroupFilterName

The name of the filter. Filter names are case-sensitive.

values :: NonEmpty Text

One or more filter values. Allowed filter values vary by group filter name, and are case-sensitive.

newGroupFilter

Arguments

:: GroupFilterName
$sel:name:GroupFilter'

-> NonEmpty Text
$sel:values:GroupFilter'

-> GroupFilter

Create a value of GroupFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GroupFilter', groupFilter_name - The name of the filter. Filter names are case-sensitive.

$sel:values:GroupFilter', groupFilter_values - One or more filter values. Allowed filter values vary by group filter name, and are case-sensitive.

groupFilter_name :: Lens' GroupFilter GroupFilterName

The name of the filter. Filter names are case-sensitive.

groupFilter_values :: Lens' GroupFilter (NonEmpty Text)

One or more filter values. Allowed filter values vary by group filter name, and are case-sensitive.

GroupIdentifier

data GroupIdentifier

The unique identifiers for a resource group.

See: newGroupIdentifier smart constructor.

Constructors

GroupIdentifier'

Fields

groupArn :: Maybe Text

The ARN of the resource group.

groupName :: Maybe Text

The name of the resource group.

newGroupIdentifier :: GroupIdentifier

Create a value of GroupIdentifier with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupArn:GroupIdentifier', groupIdentifier_groupArn - The ARN of the resource group.

$sel:groupName:GroupIdentifier', groupIdentifier_groupName - The name of the resource group.

groupIdentifier_groupArn :: Lens' GroupIdentifier (Maybe Text)

The ARN of the resource group.

groupIdentifier_groupName :: Lens' GroupIdentifier (Maybe Text)

The name of the resource group.

GroupQuery

data GroupQuery

A mapping of a query attached to a resource group that determines the AWS resources that are members of the group.

See: newGroupQuery smart constructor.

Constructors

GroupQuery'

Fields

groupName :: Text

The name of the resource group that is associated with the specified resource query.

resourceQuery :: ResourceQuery

The resource query that determines which AWS resources are members of the associated resource group.

newGroupQuery

Arguments

:: Text
$sel:groupName:GroupQuery'

-> ResourceQuery
$sel:resourceQuery:GroupQuery'

-> GroupQuery

Create a value of GroupQuery with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupName:GroupQuery', groupQuery_groupName - The name of the resource group that is associated with the specified resource query.

$sel:resourceQuery:GroupQuery', groupQuery_resourceQuery - The resource query that determines which AWS resources are members of the associated resource group.

groupQuery_groupName :: Lens' GroupQuery Text

The name of the resource group that is associated with the specified resource query.

groupQuery_resourceQuery :: Lens' GroupQuery ResourceQuery

The resource query that determines which AWS resources are members of the associated resource group.

ListGroupResourcesItem

data ListGroupResourcesItem

A structure returned by the ListGroupResources operation that contains identity and group membership status information for one of the resources in the group.

See: newListGroupResourcesItem smart constructor.

Constructors

ListGroupResourcesItem'

Fields

status :: Maybe ResourceStatus

A structure that contains the status of this resource's membership in the group.

This field is present in the response only if the group is of type AWS::EC2::HostManagement.

identifier :: Maybe ResourceIdentifier

newListGroupResourcesItem :: ListGroupResourcesItem

Create a value of ListGroupResourcesItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListGroupResourcesItem', listGroupResourcesItem_status - A structure that contains the status of this resource's membership in the group.

This field is present in the response only if the group is of type AWS::EC2::HostManagement.

$sel:identifier:ListGroupResourcesItem', listGroupResourcesItem_identifier - Undocumented member.

listGroupResourcesItem_status :: Lens' ListGroupResourcesItem (Maybe ResourceStatus)

A structure that contains the status of this resource's membership in the group.

This field is present in the response only if the group is of type AWS::EC2::HostManagement.

listGroupResourcesItem_identifier :: Lens' ListGroupResourcesItem (Maybe ResourceIdentifier)

Undocumented member.

PendingResource

data PendingResource

A structure that identifies a resource that is currently pending addition to the group as a member. Adding a resource to a resource group happens asynchronously as a background task and this one isn't completed yet.

See: newPendingResource smart constructor.

Constructors

PendingResource'

Fields

resourceArn :: Maybe Text

The Amazon resource name (ARN) of the resource that's in a pending state.

newPendingResource :: PendingResource

Create a value of PendingResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:PendingResource', pendingResource_resourceArn - The Amazon resource name (ARN) of the resource that's in a pending state.

pendingResource_resourceArn :: Lens' PendingResource (Maybe Text)

The Amazon resource name (ARN) of the resource that's in a pending state.

QueryError

data QueryError

A two-part error structure that can occur in ListGroupResources or SearchResources operations on CloudFormation stack-based queries. The error occurs if the CloudFormation stack on which the query is based either does not exist, or has a status that renders the stack inactive. A QueryError occurrence does not necessarily mean that AWS Resource Groups could not complete the operation, but the resulting group might have no member resources.

See: newQueryError smart constructor.

Constructors

QueryError'

Fields

errorCode :: Maybe QueryErrorCode

Possible values are CLOUDFORMATION_STACK_INACTIVE and CLOUDFORMATION_STACK_NOT_EXISTING.

message :: Maybe Text

A message that explains the ErrorCode value. Messages might state that the specified CloudFormation stack does not exist (or no longer exists). For CLOUDFORMATION_STACK_INACTIVE, the message typically states that the CloudFormation stack has a status that is not (or no longer) active, such as CREATE_FAILED.

newQueryError :: QueryError

Create a value of QueryError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCode:QueryError', queryError_errorCode - Possible values are CLOUDFORMATION_STACK_INACTIVE and CLOUDFORMATION_STACK_NOT_EXISTING.

$sel:message:QueryError', queryError_message - A message that explains the ErrorCode value. Messages might state that the specified CloudFormation stack does not exist (or no longer exists). For CLOUDFORMATION_STACK_INACTIVE, the message typically states that the CloudFormation stack has a status that is not (or no longer) active, such as CREATE_FAILED.

queryError_errorCode :: Lens' QueryError (Maybe QueryErrorCode)

Possible values are CLOUDFORMATION_STACK_INACTIVE and CLOUDFORMATION_STACK_NOT_EXISTING.

queryError_message :: Lens' QueryError (Maybe Text)

A message that explains the ErrorCode value. Messages might state that the specified CloudFormation stack does not exist (or no longer exists). For CLOUDFORMATION_STACK_INACTIVE, the message typically states that the CloudFormation stack has a status that is not (or no longer) active, such as CREATE_FAILED.

ResourceFilter

data ResourceFilter

A filter name and value pair that is used to obtain more specific results from a list of resources.

See: newResourceFilter smart constructor.

Constructors

ResourceFilter'

Fields

name :: ResourceFilterName

The name of the filter. Filter names are case-sensitive.

values :: NonEmpty Text

One or more filter values. Allowed filter values vary by resource filter name, and are case-sensitive.

newResourceFilter

Arguments

:: ResourceFilterName
$sel:name:ResourceFilter'

-> NonEmpty Text
$sel:values:ResourceFilter'

-> ResourceFilter

Create a value of ResourceFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ResourceFilter', resourceFilter_name - The name of the filter. Filter names are case-sensitive.

$sel:values:ResourceFilter', resourceFilter_values - One or more filter values. Allowed filter values vary by resource filter name, and are case-sensitive.

resourceFilter_name :: Lens' ResourceFilter ResourceFilterName

The name of the filter. Filter names are case-sensitive.

resourceFilter_values :: Lens' ResourceFilter (NonEmpty Text)

One or more filter values. Allowed filter values vary by resource filter name, and are case-sensitive.

ResourceIdentifier

data ResourceIdentifier

A structure that contains the ARN of a resource and its resource type.

See: newResourceIdentifier smart constructor.

Constructors

ResourceIdentifier'

Fields

resourceType :: Maybe Text

The resource type of a resource, such as AWS::EC2::Instance.

resourceArn :: Maybe Text

The ARN of a resource.

newResourceIdentifier :: ResourceIdentifier

Create a value of ResourceIdentifier with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ResourceIdentifier', resourceIdentifier_resourceType - The resource type of a resource, such as AWS::EC2::Instance.

$sel:resourceArn:ResourceIdentifier', resourceIdentifier_resourceArn - The ARN of a resource.

resourceIdentifier_resourceType :: Lens' ResourceIdentifier (Maybe Text)

The resource type of a resource, such as AWS::EC2::Instance.

resourceIdentifier_resourceArn :: Lens' ResourceIdentifier (Maybe Text)

The ARN of a resource.

ResourceQuery

data ResourceQuery

The query that is used to define a resource group or a search for resources. A query specifies both a query type and a query string as a JSON object. See the examples section for example JSON strings.

The examples that follow are shown as standard JSON strings. If you include such a string as a parameter to the AWS CLI or an SDK API, you might need to 'escape' the string into a single line. For example, see the Quoting strings in the AWS CLI User Guide.

Example 1

The following generic example shows a resource query JSON string that includes only resources that meet the following criteria:

The resource type must be either resource_type1 or resource_type2.

The resource must have a tag Key1 with a value of either ValueA or ValueB.

The resource must have a tag Key2 with a value of either ValueC or ValueD.

{ "Type": "TAG_FILTERS_1_0", "Query": { "ResourceTypeFilters": [ "resource_type1", "resource_type2"], "TagFilters": [ { "Key": "Key1", "Values": ["ValueA","ValueB"] }, { "Key":"Key2", "Values":["ValueC","ValueD"] } ] } }

This has the equivalent "shortcut" syntax of the following:

{ "Type": "TAG_FILTERS_1_0", "Query": { "ResourceTypeFilters": [ "resource_type1", "resource_type2"], "TagFilters": [ { "Key1": ["ValueA","ValueB"] }, { "Key2": ["ValueC","ValueD"] } ] } }

Example 2

The following example shows a resource query JSON string that includes only Amazon EC2 instances that are tagged Stage with a value of Test.

{ "Type": "TAG_FILTERS_1_0", "Query": "{ "ResourceTypeFilters": "AWS::EC2::Instance", "TagFilters": { "Stage": "Test" } } }

Example 3

The following example shows a resource query JSON string that includes resource of any supported type as long as it is tagged Stage with a value of Prod.

{ "Type": "TAG_FILTERS_1_0", "Query": { "ResourceTypeFilters": "AWS::AllSupported", "TagFilters": { "Stage": "Prod" } } }

Example 4

The following example shows a resource query JSON string that includes only Amazon EC2 instances and Amazon S3 buckets that are part of the specified AWS CloudFormation stack.

{ "Type": "CLOUDFORMATION_STACK_1_0", "Query": { "ResourceTypeFilters": [ "AWS::EC2::Instance", "AWS::S3::Bucket" ], "StackIdentifier": "arn:aws:cloudformation:us-west-2:123456789012:stack/AWStestuseraccount/fb0d5000-aba8-00e8-aa9e-50d5cEXAMPLE" } }

See: newResourceQuery smart constructor.

Constructors

ResourceQuery'

Fields

type' :: QueryType

The type of the query. You can use the following values:

CLOUDFORMATION_STACK_1_0: Specifies that the Query contains an ARN for a CloudFormation stack.

TAG_FILTERS_1_0: Specifies that the Query parameter contains a JSON string that represents a collection of simple tag filters for resource types and tags. The JSON string uses a syntax similar to the GetResources operation, but uses only the ResourceTypeFilters and TagFilters fields. If you specify more than one tag key, only resources that match all tag keys, and at least one value of each specified tag key, are returned in your query. If you specify more than one value for a tag key, a resource matches the filter if it has a tag key value that matches any of the specified values.

For example, consider the following sample query for resources that have two tags, Stage and Version, with two values each:

[{"Stage":["Test","Deploy"]},{"Version":["1","2"]}]

The results of this query could include the following.

An EC2 instance that has the following two tags: {"Stage":"Deploy"}, and {"Version":"2"}

An S3 bucket that has the following two tags: {"Stage":"Test"}, and {"Version":"1"}

The query would not include the following items in the results, however.

An EC2 instance that has only the following tag: {"Stage":"Deploy"}.

The instance does not have all of the tag keys specified in the filter, so it is excluded from the results.

An RDS database that has the following two tags: {"Stage":"Archived"} and {"Version":"4"}

The database has all of the tag keys, but none of those keys has an associated value that matches at least one of the specified values in the filter.

query :: Text

The query that defines a group or a search.

newResourceQuery

Arguments

:: QueryType
$sel:type':ResourceQuery'

-> Text
$sel:query:ResourceQuery'

-> ResourceQuery

Create a value of ResourceQuery with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':ResourceQuery', resourceQuery_type - The type of the query. You can use the following values:

CLOUDFORMATION_STACK_1_0: Specifies that the Query contains an ARN for a CloudFormation stack.

TAG_FILTERS_1_0: Specifies that the Query parameter contains a JSON string that represents a collection of simple tag filters for resource types and tags. The JSON string uses a syntax similar to the GetResources operation, but uses only the ResourceTypeFilters and TagFilters fields. If you specify more than one tag key, only resources that match all tag keys, and at least one value of each specified tag key, are returned in your query. If you specify more than one value for a tag key, a resource matches the filter if it has a tag key value that matches any of the specified values.

For example, consider the following sample query for resources that have two tags, Stage and Version, with two values each:

[{"Stage":["Test","Deploy"]},{"Version":["1","2"]}]

The results of this query could include the following.

An EC2 instance that has the following two tags: {"Stage":"Deploy"}, and {"Version":"2"}

An S3 bucket that has the following two tags: {"Stage":"Test"}, and {"Version":"1"}

The query would not include the following items in the results, however.

An EC2 instance that has only the following tag: {"Stage":"Deploy"}.

The instance does not have all of the tag keys specified in the filter, so it is excluded from the results.

An RDS database that has the following two tags: {"Stage":"Archived"} and {"Version":"4"}

The database has all of the tag keys, but none of those keys has an associated value that matches at least one of the specified values in the filter.

$sel:query:ResourceQuery', resourceQuery_searchQuery - The query that defines a group or a search.

resourceQuery_type :: Lens' ResourceQuery QueryType

The type of the query. You can use the following values:

CLOUDFORMATION_STACK_1_0: Specifies that the Query contains an ARN for a CloudFormation stack.

TAG_FILTERS_1_0: Specifies that the Query parameter contains a JSON string that represents a collection of simple tag filters for resource types and tags. The JSON string uses a syntax similar to the GetResources operation, but uses only the ResourceTypeFilters and TagFilters fields. If you specify more than one tag key, only resources that match all tag keys, and at least one value of each specified tag key, are returned in your query. If you specify more than one value for a tag key, a resource matches the filter if it has a tag key value that matches any of the specified values.

For example, consider the following sample query for resources that have two tags, Stage and Version, with two values each:

[{"Stage":["Test","Deploy"]},{"Version":["1","2"]}]

The results of this query could include the following.

An EC2 instance that has the following two tags: {"Stage":"Deploy"}, and {"Version":"2"}

An S3 bucket that has the following two tags: {"Stage":"Test"}, and {"Version":"1"}

The query would not include the following items in the results, however.

An EC2 instance that has only the following tag: {"Stage":"Deploy"}.

The instance does not have all of the tag keys specified in the filter, so it is excluded from the results.

An RDS database that has the following two tags: {"Stage":"Archived"} and {"Version":"4"}

The database has all of the tag keys, but none of those keys has an associated value that matches at least one of the specified values in the filter.

resourceQuery_searchQuery :: Lens' ResourceQuery Text

The query that defines a group or a search.

ResourceStatus

data ResourceStatus

A structure that identifies the current group membership status for a resource. Adding a resource to a resource group is performed asynchronously as a background task. A PENDING status indicates, for this resource, that the process isn't completed yet.

See: newResourceStatus smart constructor.

Constructors

ResourceStatus'

Fields

name :: Maybe ResourceStatusValue

The current status.

newResourceStatus :: ResourceStatus

Create a value of ResourceStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ResourceStatus', resourceStatus_name - The current status.

resourceStatus_name :: Lens' ResourceStatus (Maybe ResourceStatusValue)

The current status.