Network.AWS.ResourceGroups.Types.GroupConfiguration

Description

Documentation

data GroupConfiguration

A service configuration associated with a resource group. The configuration options are determined by the AWS service that defines the Type, and specifies which resources can be included in the group. You can add a service configuration when you create the group by using CreateGroup, or later by using the PutGroupConfiguration operation. For details about group service configuration syntax, see Service configurations for resource groups.

See: newGroupConfiguration smart constructor.

Constructors

GroupConfiguration'

Fields

status :: Maybe GroupConfigurationStatus

The current status of an attempt to update the group configuration.

configuration :: Maybe [GroupConfigurationItem]

The configuration currently associated with the group and in effect.

failureReason :: Maybe Text

If present, the reason why a request to update the group configuration failed.

proposedConfiguration :: Maybe [GroupConfigurationItem]

If present, the new configuration that is in the process of being applied to the group.

newGroupConfiguration :: GroupConfiguration

Create a value of GroupConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GroupConfiguration', groupConfiguration_status - The current status of an attempt to update the group configuration.

$sel:configuration:GroupConfiguration', groupConfiguration_configuration - The configuration currently associated with the group and in effect.

$sel:failureReason:GroupConfiguration', groupConfiguration_failureReason - If present, the reason why a request to update the group configuration failed.

$sel:proposedConfiguration:GroupConfiguration', groupConfiguration_proposedConfiguration - If present, the new configuration that is in the process of being applied to the group.

groupConfiguration_status :: Lens' GroupConfiguration (Maybe GroupConfigurationStatus)

The current status of an attempt to update the group configuration.

groupConfiguration_configuration :: Lens' GroupConfiguration (Maybe [GroupConfigurationItem])

The configuration currently associated with the group and in effect.

groupConfiguration_failureReason :: Lens' GroupConfiguration (Maybe Text)

If present, the reason why a request to update the group configuration failed.

groupConfiguration_proposedConfiguration :: Lens' GroupConfiguration (Maybe [GroupConfigurationItem])

If present, the new configuration that is in the process of being applied to the group.