Network.AWS.ResourceGroups.UngroupResources

Description

Removes the specified resources from the specified group.

Minimum permissions

To run this command, you must have the following permissions:

resource-groups:UngroupResources

Creating a Request

data UngroupResources

See: newUngroupResources smart constructor.

Constructors

UngroupResources'

Fields

group' :: Text

The name or the ARN of the resource group from which to remove the resources.

resourceArns :: NonEmpty Text

The ARNs of the resources to be removed from the group.

newUngroupResources

Arguments

:: Text
$sel:group':UngroupResources'

-> NonEmpty Text
$sel:resourceArns:UngroupResources'

-> UngroupResources

Create a value of UngroupResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:group':UngroupResources', ungroupResources_group - The name or the ARN of the resource group from which to remove the resources.

$sel:resourceArns:UngroupResources', ungroupResources_resourceArns - The ARNs of the resources to be removed from the group.

Request Lenses

ungroupResources_group :: Lens' UngroupResources Text

The name or the ARN of the resource group from which to remove the resources.

ungroupResources_resourceArns :: Lens' UngroupResources (NonEmpty Text)

The ARNs of the resources to be removed from the group.

Destructuring the Response

data UngroupResourcesResponse

See: newUngroupResourcesResponse smart constructor.

Constructors

UngroupResourcesResponse'

Fields

pending :: Maybe [PendingResource]

A list of any resources that are still in the process of being removed from the group by this operation. These pending removals continue asynchronously. You can check the status of pending removals by using the ListGroupResources operation. After the resource is successfully removed, it no longer appears in the response.

succeeded :: Maybe (NonEmpty Text)

A list of resources that were successfully removed from the group by this operation.

failed :: Maybe [FailedResource]

A list of any resources that failed to be removed from the group by this operation.

httpStatus :: Int

The response's http status code.

newUngroupResourcesResponse

Arguments

:: Int
$sel:httpStatus:UngroupResourcesResponse'

-> UngroupResourcesResponse

Create a value of UngroupResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pending:UngroupResourcesResponse', ungroupResourcesResponse_pending - A list of any resources that are still in the process of being removed from the group by this operation. These pending removals continue asynchronously. You can check the status of pending removals by using the ListGroupResources operation. After the resource is successfully removed, it no longer appears in the response.

$sel:succeeded:UngroupResourcesResponse', ungroupResourcesResponse_succeeded - A list of resources that were successfully removed from the group by this operation.

$sel:failed:UngroupResourcesResponse', ungroupResourcesResponse_failed - A list of any resources that failed to be removed from the group by this operation.

$sel:httpStatus:UngroupResourcesResponse', ungroupResourcesResponse_httpStatus - The response's http status code.

Response Lenses

ungroupResourcesResponse_pending :: Lens' UngroupResourcesResponse (Maybe [PendingResource])

A list of any resources that are still in the process of being removed from the group by this operation. These pending removals continue asynchronously. You can check the status of pending removals by using the ListGroupResources operation. After the resource is successfully removed, it no longer appears in the response.

ungroupResourcesResponse_succeeded :: Lens' UngroupResourcesResponse (Maybe (NonEmpty Text))

A list of resources that were successfully removed from the group by this operation.

ungroupResourcesResponse_failed :: Lens' UngroupResourcesResponse (Maybe [FailedResource])

A list of any resources that failed to be removed from the group by this operation.

ungroupResourcesResponse_httpStatus :: Lens' UngroupResourcesResponse Int

The response's http status code.