Network.AWS.ResourceGroups.Untag

Description

Deletes tags from a specified resource group.

Minimum permissions

To run this command, you must have the following permissions:

resource-groups:Untag

Creating a Request

data Untag

See: newUntag smart constructor.

Constructors

Untag'

Fields

arn :: Text

The ARN of the resource group from which to remove tags. The command removed both the specified keys and any values associated with those keys.

keys :: [Text]

The keys of the tags to be removed.

newUntag

Arguments

:: Text
$sel:arn:Untag'

-> Untag

Create a value of Untag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Untag', untag_arn - The ARN of the resource group from which to remove tags. The command removed both the specified keys and any values associated with those keys.

$sel:keys:Untag', untag_keys - The keys of the tags to be removed.

Request Lenses

untag_arn :: Lens' Untag Text

The ARN of the resource group from which to remove tags. The command removed both the specified keys and any values associated with those keys.

untag_keys :: Lens' Untag [Text]

The keys of the tags to be removed.

Destructuring the Response

data UntagResponse

See: newUntagResponse smart constructor.

Constructors

UntagResponse'

Fields

arn :: Maybe Text

The ARN of the resource group from which tags have been removed.

keys :: Maybe [Text]

The keys of the tags that were removed.

httpStatus :: Int

The response's http status code.

newUntagResponse

Arguments

:: Int
$sel:httpStatus:UntagResponse'

-> UntagResponse

Create a value of UntagResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Untag', untagResponse_arn - The ARN of the resource group from which tags have been removed.

$sel:keys:Untag', untagResponse_keys - The keys of the tags that were removed.

$sel:httpStatus:UntagResponse', untagResponse_httpStatus - The response's http status code.

Response Lenses

untagResponse_arn :: Lens' UntagResponse (Maybe Text)

The ARN of the resource group from which tags have been removed.

untagResponse_keys :: Lens' UntagResponse (Maybe [Text])

The keys of the tags that were removed.

untagResponse_httpStatus :: Lens' UntagResponse Int

The response's http status code.