Network.AWS.ResourceGroups.UpdateGroupQuery

Description

Updates the resource query of a group. For more information about resource queries, see Create a tag-based group in Resource Groups.

Minimum permissions

To run this command, you must have the following permissions:

resource-groups:UpdateGroupQuery

Creating a Request

data UpdateGroupQuery

See: newUpdateGroupQuery smart constructor.

Constructors

UpdateGroupQuery'

Fields

groupName :: Maybe Text

Don't use this parameter. Use Group instead.

group' :: Maybe Text

The name or the ARN of the resource group to query.

resourceQuery :: ResourceQuery

The resource query to determine which AWS resources are members of this resource group.

A resource group can contain either a Configuration or a ResourceQuery, but not both.

newUpdateGroupQuery

Arguments

:: ResourceQuery
$sel:resourceQuery:UpdateGroupQuery'

-> UpdateGroupQuery

Create a value of UpdateGroupQuery with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupName:UpdateGroupQuery', updateGroupQuery_groupName - Don't use this parameter. Use Group instead.

$sel:group':UpdateGroupQuery', updateGroupQuery_group - The name or the ARN of the resource group to query.

$sel:resourceQuery:UpdateGroupQuery', updateGroupQuery_resourceQuery - The resource query to determine which AWS resources are members of this resource group.

A resource group can contain either a Configuration or a ResourceQuery, but not both.

Request Lenses

updateGroupQuery_groupName :: Lens' UpdateGroupQuery (Maybe Text)

Don't use this parameter. Use Group instead.

updateGroupQuery_group :: Lens' UpdateGroupQuery (Maybe Text)

The name or the ARN of the resource group to query.

updateGroupQuery_resourceQuery :: Lens' UpdateGroupQuery ResourceQuery

The resource query to determine which AWS resources are members of this resource group.

A resource group can contain either a Configuration or a ResourceQuery, but not both.

Destructuring the Response

data UpdateGroupQueryResponse

See: newUpdateGroupQueryResponse smart constructor.

Constructors

UpdateGroupQueryResponse'

Fields

groupQuery :: Maybe GroupQuery

The updated resource query associated with the resource group after the update.

httpStatus :: Int

The response's http status code.

newUpdateGroupQueryResponse

Arguments

:: Int
$sel:httpStatus:UpdateGroupQueryResponse'

-> UpdateGroupQueryResponse

Create a value of UpdateGroupQueryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupQuery:UpdateGroupQueryResponse', updateGroupQueryResponse_groupQuery - The updated resource query associated with the resource group after the update.

$sel:httpStatus:UpdateGroupQueryResponse', updateGroupQueryResponse_httpStatus - The response's http status code.

Response Lenses

updateGroupQueryResponse_groupQuery :: Lens' UpdateGroupQueryResponse (Maybe GroupQuery)

The updated resource query associated with the resource group after the update.

updateGroupQueryResponse_httpStatus :: Lens' UpdateGroupQueryResponse Int

The response's http status code.