Network.AWS.ResourceGroupsTagging

Description

Derived from API version 2017-01-26 of the AWS service descriptions, licensed under Apache 2.0.

Resource Groups Tagging API

Service Configuration

defaultService :: Service

API version 2017-01-26 of the Amazon Resource Groups Tagging API SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by ResourceGroupsTagging.

InvalidParameterException

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

This error indicates one of the following:

A parameter is missing.

A malformed string was supplied for the request parameter.

An out-of-range value was supplied for the request parameter.

The target ID is invalid, unsupported, or doesn't exist.

You can't access the Amazon S3 bucket for report storage. For more information, see Additional Requirements for Organization-wide Tag Compliance Reports in the AWS Organizations User Guide.

ConstraintViolationException

_ConstraintViolationException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied because performing this operation violates a constraint.

Some of the reasons in the following list might not apply to this specific operation.

You must meet the prerequisites for using tag policies. For information, see Prerequisites and Permissions for Using Tag Policies in the AWS Organizations User Guide.

You must enable the tag policies service principal (tagpolicies.tag.amazonaws.com) to integrate with AWS Organizations For information, see EnableAWSServiceAccess.

You must have a tag policy attached to the organization root, an OU, or an account.

ThrottledException

_ThrottledException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied to limit the frequency of submitted requests.

PaginationTokenExpiredException

_PaginationTokenExpiredException :: AsError a => Getting (First ServiceError) a ServiceError

A PaginationToken is valid for a maximum of 15 minutes. Your request was denied because the specified PaginationToken has expired.

ConcurrentModificationException

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

The target of the operation is currently being modified by a different request. Try again later.

InternalServiceException

_InternalServiceException :: AsError a => Getting (First ServiceError) a ServiceError

The request processing failed because of an unknown error, exception, or failure. You can retry the request.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetComplianceSummary (Paginated)

data GetComplianceSummary

See: newGetComplianceSummary smart constructor.

Constructors

GetComplianceSummary' (Maybe [GroupByAttribute]) (Maybe Text) (Maybe (NonEmpty Text)) (Maybe [Text]) (Maybe (NonEmpty Text)) (Maybe (NonEmpty Text)) (Maybe Natural)

newGetComplianceSummary :: GetComplianceSummary

Create a value of GetComplianceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupBy:GetComplianceSummary', getComplianceSummary_groupBy - Specifies a list of attributes to group the counts of noncompliant resources by. If supplied, the counts are sorted by those attributes.

$sel:paginationToken:GetComplianceSummary', getComplianceSummary_paginationToken - Specifies a PaginationToken response value from a previous request to indicate that you want the next page of results. Leave this parameter empty in your initial request.

$sel:targetIdFilters:GetComplianceSummary', getComplianceSummary_targetIdFilters - Specifies target identifiers (usually, specific account IDs) to limit the output by. If you use this parameter, the count of returned noncompliant resources includes only resources with the specified target IDs.

$sel:resourceTypeFilters:GetComplianceSummary', getComplianceSummary_resourceTypeFilters - Specifies that you want the response to include information for only resources of the specified types. The format of each resource type is service[:resourceType]. For example, specifying a resource type of ec2 returns all Amazon EC2 resources (which includes EC2 instances). Specifying a resource type of ec2:instance returns only EC2 instances.

The string for each service name and resource type is the same as that embedded in a resource's Amazon Resource Name (ARN). Consult the /AWS General Reference/ for the following:

For a list of service name strings, see AWS Service Namespaces.

For resource type strings, see Example ARNs.

For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

You can specify multiple resource types by using a comma separated array. The array can include up to 100 items. Note that the length constraint requirement applies to each resource type filter.

$sel:regionFilters:GetComplianceSummary', getComplianceSummary_regionFilters - Specifies a list of AWS Regions to limit the output by. If you use this parameter, the count of returned noncompliant resources includes only resources in the specified Regions.

$sel:tagKeyFilters:GetComplianceSummary', getComplianceSummary_tagKeyFilters - Specifies that you want the response to include information for only resources that have tags with the specified tag keys. If you use this parameter, the count of returned noncompliant resources includes only resources that have the specified tag keys.

$sel:maxResults:GetComplianceSummary', getComplianceSummary_maxResults - Specifies the maximum number of results to be returned in each page. A query can return fewer than this maximum, even if there are more results still to return. You should always check the PaginationToken response value to see if there are more results. You can specify a minimum of 1 and a maximum value of 100.

data GetComplianceSummaryResponse

See: newGetComplianceSummaryResponse smart constructor.

Constructors

GetComplianceSummaryResponse' (Maybe Text) (Maybe [Summary]) Int

newGetComplianceSummaryResponse

Arguments

:: Int
$sel:httpStatus:GetComplianceSummaryResponse'

-> GetComplianceSummaryResponse

Create a value of GetComplianceSummaryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:paginationToken:GetComplianceSummary', getComplianceSummaryResponse_paginationToken - A string that indicates that there is more data available than this response contains. To receive the next part of the response, specify this response value as the PaginationToken value in the request for the next page.

$sel:summaryList:GetComplianceSummaryResponse', getComplianceSummaryResponse_summaryList - A table that shows counts of noncompliant resources.

$sel:httpStatus:GetComplianceSummaryResponse', getComplianceSummaryResponse_httpStatus - The response's http status code.

DescribeReportCreation

data DescribeReportCreation

See: newDescribeReportCreation smart constructor.

Constructors

DescribeReportCreation'

newDescribeReportCreation :: DescribeReportCreation

Create a value of DescribeReportCreation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

data DescribeReportCreationResponse

See: newDescribeReportCreationResponse smart constructor.

Constructors

DescribeReportCreationResponse' (Maybe Text) (Maybe Text) (Maybe Text) Int

newDescribeReportCreationResponse

Arguments

:: Int
$sel:httpStatus:DescribeReportCreationResponse'

-> DescribeReportCreationResponse

Create a value of DescribeReportCreationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeReportCreationResponse', describeReportCreationResponse_status - Reports the status of the operation.

The operation status can be one of the following:

RUNNING - Report creation is in progress.

SUCCEEDED - Report creation is complete. You can open the report from the Amazon S3 bucket that you specified when you ran StartReportCreation.

FAILED - Report creation timed out or the Amazon S3 bucket is not accessible.

NO REPORT - No report was generated in the last 90 days.

$sel:s3Location:DescribeReportCreationResponse', describeReportCreationResponse_s3Location - The path to the Amazon S3 bucket where the report was stored on creation.

$sel:errorMessage:DescribeReportCreationResponse', describeReportCreationResponse_errorMessage - Details of the common errors that all operations return.

$sel:httpStatus:DescribeReportCreationResponse', describeReportCreationResponse_httpStatus - The response's http status code.

GetTagKeys (Paginated)

data GetTagKeys

See: newGetTagKeys smart constructor.

Constructors

GetTagKeys' (Maybe Text)

newGetTagKeys :: GetTagKeys

Create a value of GetTagKeys with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:paginationToken:GetTagKeys', getTagKeys_paginationToken - Specifies a PaginationToken response value from a previous request to indicate that you want the next page of results. Leave this parameter empty in your initial request.

data GetTagKeysResponse

See: newGetTagKeysResponse smart constructor.

Constructors

GetTagKeysResponse' (Maybe Text) (Maybe [Text]) Int

newGetTagKeysResponse

Arguments

:: Int
$sel:httpStatus:GetTagKeysResponse'

-> GetTagKeysResponse

Create a value of GetTagKeysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:paginationToken:GetTagKeys', getTagKeysResponse_paginationToken - A string that indicates that there is more data available than this response contains. To receive the next part of the response, specify this response value as the PaginationToken value in the request for the next page.

$sel:tagKeys:GetTagKeysResponse', getTagKeysResponse_tagKeys - A list of all tag keys in the AWS account.

$sel:httpStatus:GetTagKeysResponse', getTagKeysResponse_httpStatus - The response's http status code.

TagResources

data TagResources

See: newTagResources smart constructor.

Constructors

TagResources' (NonEmpty Text) (HashMap Text Text)

newTagResources

Arguments

:: NonEmpty Text
$sel:resourceARNList:TagResources'

-> TagResources

Create a value of TagResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceARNList:TagResources', tagResources_resourceARNList - Specifies the list of ARNs of the resources that you want to apply tags to.

An ARN (Amazon Resource Name) uniquely identifies a resource. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

$sel:tags:TagResources', tagResources_tags - Specifies a list of tags that you want to add to the specified resources. A tag consists of a key and a value that you define.

data TagResourcesResponse

See: newTagResourcesResponse smart constructor.

Constructors

TagResourcesResponse' (Maybe (HashMap Text FailureInfo)) Int

newTagResourcesResponse

Arguments

:: Int
$sel:httpStatus:TagResourcesResponse'

-> TagResourcesResponse

Create a value of TagResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedResourcesMap:TagResourcesResponse', tagResourcesResponse_failedResourcesMap - A map containing a key-value pair for each failed item that couldn't be tagged. The key is the ARN of the failed resource. The value is a FailureInfo object that contains an error code, a status code, and an error message. If there are no errors, the FailedResourcesMap is empty.

$sel:httpStatus:TagResourcesResponse', tagResourcesResponse_httpStatus - The response's http status code.

GetTagValues (Paginated)

data GetTagValues

See: newGetTagValues smart constructor.

Constructors

GetTagValues' (Maybe Text) Text

newGetTagValues

Arguments

:: Text
$sel:key:GetTagValues'

-> GetTagValues

Create a value of GetTagValues with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:paginationToken:GetTagValues', getTagValues_paginationToken - Specifies a PaginationToken response value from a previous request to indicate that you want the next page of results. Leave this parameter empty in your initial request.

$sel:key:GetTagValues', getTagValues_key - Specifies the tag key for which you want to list all existing values that are currently used in the specified AWS Region for the calling AWS account.

data GetTagValuesResponse

See: newGetTagValuesResponse smart constructor.

Constructors

GetTagValuesResponse' (Maybe Text) (Maybe [Text]) Int

newGetTagValuesResponse

Arguments

:: Int
$sel:httpStatus:GetTagValuesResponse'

-> GetTagValuesResponse

Create a value of GetTagValuesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:paginationToken:GetTagValues', getTagValuesResponse_paginationToken - A string that indicates that there is more data available than this response contains. To receive the next part of the response, specify this response value as the PaginationToken value in the request for the next page.

$sel:tagValues:GetTagValuesResponse', getTagValuesResponse_tagValues - A list of all tag values for the specified key currently used in the specified AWS Region for the calling AWS account.

$sel:httpStatus:GetTagValuesResponse', getTagValuesResponse_httpStatus - The response's http status code.

StartReportCreation

data StartReportCreation

See: newStartReportCreation smart constructor.

Constructors

StartReportCreation' Text

newStartReportCreation

Arguments

:: Text
$sel:s3Bucket:StartReportCreation'

-> StartReportCreation

Create a value of StartReportCreation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Bucket:StartReportCreation', startReportCreation_s3Bucket - The name of the Amazon S3 bucket where the report will be stored; for example:

awsexamplebucket

For more information on S3 bucket requirements, including an example bucket policy, see the example S3 bucket policy on this page.

data StartReportCreationResponse

See: newStartReportCreationResponse smart constructor.

Constructors

StartReportCreationResponse' Int

newStartReportCreationResponse

Arguments

:: Int
$sel:httpStatus:StartReportCreationResponse'

-> StartReportCreationResponse

Create a value of StartReportCreationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StartReportCreationResponse', startReportCreationResponse_httpStatus - The response's http status code.

GetResources (Paginated)

data GetResources

See: newGetResources smart constructor.

Constructors

GetResources' (Maybe Text) (Maybe Int) (Maybe Bool) (Maybe Bool) (Maybe [Text]) (Maybe [TagFilter]) (Maybe (NonEmpty Text)) (Maybe Int)

newGetResources :: GetResources

Create a value of GetResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:paginationToken:GetResources', getResources_paginationToken - Specifies a PaginationToken response value from a previous request to indicate that you want the next page of results. Leave this parameter empty in your initial request.

$sel:resourcesPerPage:GetResources', getResources_resourcesPerPage - Specifies the maximum number of results to be returned in each page. A query can return fewer than this maximum, even if there are more results still to return. You should always check the PaginationToken response value to see if there are more results. You can specify a minimum of 1 and a maximum value of 100.

$sel:excludeCompliantResources:GetResources', getResources_excludeCompliantResources - Specifies whether to exclude resources that are compliant with the tag policy. Set this to true if you are interested in retrieving information on noncompliant resources only.

You can use this parameter only if the IncludeComplianceDetails parameter is also set to true.

$sel:includeComplianceDetails:GetResources', getResources_includeComplianceDetails - Specifies whether to include details regarding the compliance with the effective tag policy. Set this to true to determine whether resources are compliant with the tag policy and to get details.

$sel:resourceTypeFilters:GetResources', getResources_resourceTypeFilters - Specifies the resource types that you want included in the response. The format of each resource type is service[:resourceType]. For example, specifying a resource type of ec2 returns all Amazon EC2 resources (which includes EC2 instances). Specifying a resource type of ec2:instance returns only EC2 instances.

The string for each service name and resource type is the same as that embedded in a resource's Amazon Resource Name (ARN). Consult the /AWS General Reference/ for the following:

For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

You can specify multiple resource types by using an array. The array can include up to 100 items. Note that the length constraint requirement applies to each resource type filter.

$sel:tagFilters:GetResources', getResources_tagFilters - Specifies a list of TagFilters (keys and values) to restrict the output to only those resources that have the specified tag and, if included, the specified value. Each TagFilter must contain a key with values optional. A request can include up to 50 keys, and each key can include up to 20 values.

Note the following when deciding how to use TagFilters:

If you don't specify a TagFilter, the response includes all resources that are currently tagged or ever had a tag. Resources that currently don't have tags are shown with an empty tag set, like this: "Tags": [].

If you specify more than one filter in a single request, the response returns only those resources that satisfy all filters.

If you specify a filter that contains more than one value for a key, the response returns resources that match any of the specified values for that key.

If you don't specify any values for a key, the response returns resources that are tagged with that key and any or no value.

For example, for the following filters: filter1= {keyA,{value1}}, filter2={keyB,{value2,value3,value4}}, filter3= {keyC}:

GetResources({filter1}) returns resources tagged with key1=value1

GetResources({filter2}) returns resources tagged with key2=value2 or key2=value3 or key2=value4

GetResources({filter3}) returns resources tagged with any tag with the key key3, and with any or no value

GetResources({filter1,filter2,filter3}) returns resources tagged with (key1=value1) and (key2=value2 or key2=value3 or key2=value4) and (key3, any or no value)

$sel:resourceARNList:GetResources', getResources_resourceARNList - Specifies a list of ARNs of resources for which you want to retrieve tag data. You can't specify both this parameter and any of the pagination parameters (ResourcesPerPage, TagsPerPage, PaginationToken) in the same request. If you specify both, you get an Invalid Parameter exception.

If a resource specified by this parameter doesn't exist, it doesn't generate an error; it simply isn't included in the response.

An ARN (Amazon Resource Name) uniquely identifies a resource. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

$sel:tagsPerPage:GetResources', getResources_tagsPerPage - AWS recommends using ResourcesPerPage instead of this parameter.

A limit that restricts the number of tags (key and value pairs) returned by GetResources in paginated output. A resource with no tags is counted as having one tag (one key and value pair).

GetResources does not split a resource and its associated tags across pages. If the specified TagsPerPage would cause such a break, a PaginationToken is returned in place of the affected resource and its tags. Use that token in another request to get the remaining data. For example, if you specify a TagsPerPage of 100 and the account has 22 resources with 10 tags each (meaning that each resource has 10 key and value pairs), the output will consist of three pages. The first page displays the first 10 resources, each with its 10 tags. The second page displays the next 10 resources, each with its 10 tags. The third page displays the remaining 2 resources, each with its 10 tags.

You can set TagsPerPage to a minimum of 100 items up to a maximum of 500 items.

data GetResourcesResponse

See: newGetResourcesResponse smart constructor.

Constructors

GetResourcesResponse' (Maybe Text) (Maybe [ResourceTagMapping]) Int

newGetResourcesResponse

Arguments

:: Int
$sel:httpStatus:GetResourcesResponse'

-> GetResourcesResponse

Create a value of GetResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:paginationToken:GetResources', getResourcesResponse_paginationToken - A string that indicates that there is more data available than this response contains. To receive the next part of the response, specify this response value as the PaginationToken value in the request for the next page.

$sel:resourceTagMappingList:GetResourcesResponse', getResourcesResponse_resourceTagMappingList - A list of resource ARNs and the tags (keys and values) associated with those ARNs.

$sel:httpStatus:GetResourcesResponse', getResourcesResponse_httpStatus - The response's http status code.

UntagResources

data UntagResources

See: newUntagResources smart constructor.

Constructors

UntagResources' (NonEmpty Text) (NonEmpty Text)

newUntagResources

Arguments

:: NonEmpty Text
$sel:resourceARNList:UntagResources'

-> NonEmpty Text
$sel:tagKeys:UntagResources'

-> UntagResources

Create a value of UntagResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceARNList:UntagResources', untagResources_resourceARNList - Specifies a list of ARNs of the resources that you want to remove tags from.

An ARN (Amazon Resource Name) uniquely identifies a resource. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

$sel:tagKeys:UntagResources', untagResources_tagKeys - Specifies a list of tag keys that you want to remove from the specified resources.

data UntagResourcesResponse

See: newUntagResourcesResponse smart constructor.

Constructors

UntagResourcesResponse' (Maybe (HashMap Text FailureInfo)) Int

newUntagResourcesResponse

Arguments

:: Int
$sel:httpStatus:UntagResourcesResponse'

-> UntagResourcesResponse

Create a value of UntagResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedResourcesMap:UntagResourcesResponse', untagResourcesResponse_failedResourcesMap - A map containing a key-value pair for each failed item that couldn't be untagged. The key is the ARN of the failed resource. The value is a FailureInfo object that contains an error code, a status code, and an error message. If there are no errors, the FailedResourcesMap is empty.

$sel:httpStatus:UntagResourcesResponse', untagResourcesResponse_httpStatus - The response's http status code.

Types

GroupByAttribute

newtype GroupByAttribute

Constructors

GroupByAttribute'

Fields

fromGroupByAttribute :: Text

Bundled Patterns

pattern GroupByAttribute_REGION :: GroupByAttribute

pattern GroupByAttribute_RESOURCE_TYPE :: GroupByAttribute

pattern GroupByAttribute_TARGET_ID :: GroupByAttribute

ResourceErrorCode

newtype ResourceErrorCode

Constructors

ResourceErrorCode'

Fields

fromResourceErrorCode :: Text

Bundled Patterns

pattern ResourceErrorCode_InternalServiceException :: ResourceErrorCode

pattern ResourceErrorCode_InvalidParameterException :: ResourceErrorCode

TargetIdType

newtype TargetIdType

Constructors

TargetIdType'

Fields

fromTargetIdType :: Text

Bundled Patterns

pattern TargetIdType_ACCOUNT :: TargetIdType

pattern TargetIdType_OU :: TargetIdType

pattern TargetIdType_ROOT :: TargetIdType

ComplianceDetails

data ComplianceDetails

Information that shows whether a resource is compliant with the effective tag policy, including details on any noncompliant tag keys.

See: newComplianceDetails smart constructor.

Constructors

ComplianceDetails' (Maybe [Text]) (Maybe Bool) (Maybe [Text])

newComplianceDetails :: ComplianceDetails

Create a value of ComplianceDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keysWithNoncompliantValues:ComplianceDetails', complianceDetails_keysWithNoncompliantValues - These are keys defined in the effective policy that are on the resource with either incorrect case treatment or noncompliant values.

$sel:complianceStatus:ComplianceDetails', complianceDetails_complianceStatus - Whether a resource is compliant with the effective tag policy.

$sel:noncompliantKeys:ComplianceDetails', complianceDetails_noncompliantKeys - These tag keys on the resource are noncompliant with the effective tag policy.

FailureInfo

data FailureInfo

Information about the errors that are returned for each failed resource. This information can include InternalServiceException and InvalidParameterException errors. It can also include any valid error code returned by the AWS service that hosts the resource that the ARN key represents.

The following are common error codes that you might receive from other AWS services:

InternalServiceException – This can mean that the Resource Groups Tagging API didn't receive a response from another AWS service. It can also mean the the resource type in the request is not supported by the Resource Groups Tagging API. In these cases, it's safe to retry the request and then call GetResources to verify the changes.

AccessDeniedException – This can mean that you need permission to calling tagging operations in the AWS service that contains the resource. For example, to use the Resource Groups Tagging API to tag a CloudWatch alarm resource, you need permission to call TagResources and TagResource in the CloudWatch API.

For more information on errors that are generated from other AWS services, see the documentation for that service.

See: newFailureInfo smart constructor.

Constructors

FailureInfo' (Maybe ResourceErrorCode) (Maybe Text) (Maybe Int)

newFailureInfo :: FailureInfo

Create a value of FailureInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCode:FailureInfo', failureInfo_errorCode - The code of the common error. Valid values include InternalServiceException, InvalidParameterException, and any valid error code returned by the AWS service that hosts the resource that you want to tag.

$sel:errorMessage:FailureInfo', failureInfo_errorMessage - The message of the common error.

$sel:statusCode:FailureInfo', failureInfo_statusCode - The HTTP status code of the common error.

ResourceTagMapping

data ResourceTagMapping

A list of resource ARNs and the tags (keys and values) that are associated with each.

See: newResourceTagMapping smart constructor.

Constructors

ResourceTagMapping' (Maybe ComplianceDetails) (Maybe Text) (Maybe [Tag])

newResourceTagMapping :: ResourceTagMapping

Create a value of ResourceTagMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:complianceDetails:ResourceTagMapping', resourceTagMapping_complianceDetails - Information that shows whether a resource is compliant with the effective tag policy, including details on any noncompliant tag keys.

$sel:resourceARN:ResourceTagMapping', resourceTagMapping_resourceARN - The ARN of the resource.

$sel:tags:ResourceTagMapping', resourceTagMapping_tags - The tags that have been applied to one or more AWS resources.

Summary

data Summary

A count of noncompliant resources.

See: newSummary smart constructor.

Constructors

Summary' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Integer) (Maybe TargetIdType) (Maybe Text)

newSummary :: Summary

Create a value of Summary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetId:Summary', summary_targetId - The account identifier or the root identifier of the organization. If you don't know the root ID, you can call the AWS Organizations ListRoots API.

$sel:lastUpdated:Summary', summary_lastUpdated - The timestamp that shows when this summary was generated in this Region.

$sel:resourceType:Summary', summary_resourceType - The AWS resource type.

$sel:nonCompliantResources:Summary', summary_nonCompliantResources - The count of noncompliant resources.

$sel:targetIdType:Summary', summary_targetIdType - Whether the target is an account, an OU, or the organization root.

$sel:region:Summary', summary_region - The AWS Region that the summary applies to.

Tag

data Tag

The metadata that you apply to AWS resources to help you categorize and organize them. Each tag consists of a key and a value, both of which you define. For more information, see Tagging AWS Resources in the AWS General Reference.

See: newTag smart constructor.

Constructors

Tag' Text Text

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - One part of a key-value pair that makes up a tag. A key is a general label that acts like a category for more specific tag values.

$sel:value:Tag', tag_value - One part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key). The value can be empty or null.

TagFilter

data TagFilter

A list of tags (keys and values) that are used to specify the associated resources.

See: newTagFilter smart constructor.

Constructors

TagFilter' (Maybe [Text]) (Maybe Text)

newTagFilter :: TagFilter

Create a value of TagFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:TagFilter', tagFilter_values - One part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key). The value can be empty or null.

$sel:key:TagFilter', tagFilter_key - One part of a key-value pair that makes up a tag. A key is a general label that acts like a category for more specific tag values.