Network.AWS.ResourceGroupsTagging.GetComplianceSummary

Description

Returns a table that shows counts of resources that are noncompliant with their tag policies.

For more information on tag policies, see Tag Policies in the AWS Organizations User Guide.

You can call this operation only from the organization's management account and from the us-east-1 Region.

This operation supports pagination, where the response can be sent in multiple pages. You should check the PaginationToken response parameter to determine if there are additional results available to return. Repeat the query, passing the PaginationToken response parameter value as an input to the next request until you recieve a null value. A null value for PaginationToken indicates that there are no more results waiting to be returned.

This operation returns paginated results.

Creating a Request

data GetComplianceSummary

See: newGetComplianceSummary smart constructor.

Constructors

GetComplianceSummary'

Fields

maxResults :: Maybe Natural

Specifies the maximum number of results to be returned in each page. A query can return fewer than this maximum, even if there are more results still to return. You should always check the PaginationToken response value to see if there are more results. You can specify a minimum of 1 and a maximum value of 100.

regionFilters :: Maybe (NonEmpty Text)

Specifies a list of AWS Regions to limit the output by. If you use this parameter, the count of returned noncompliant resources includes only resources in the specified Regions.

paginationToken :: Maybe Text

Specifies a PaginationToken response value from a previous request to indicate that you want the next page of results. Leave this parameter empty in your initial request.

targetIdFilters :: Maybe (NonEmpty Text)

Specifies target identifiers (usually, specific account IDs) to limit the output by. If you use this parameter, the count of returned noncompliant resources includes only resources with the specified target IDs.

groupBy :: Maybe [GroupByAttribute]

Specifies a list of attributes to group the counts of noncompliant resources by. If supplied, the counts are sorted by those attributes.

resourceTypeFilters :: Maybe [Text]

Specifies that you want the response to include information for only resources of the specified types. The format of each resource type is service[:resourceType]. For example, specifying a resource type of ec2 returns all Amazon EC2 resources (which includes EC2 instances). Specifying a resource type of ec2:instance returns only EC2 instances.

The string for each service name and resource type is the same as that embedded in a resource's Amazon Resource Name (ARN). Consult the /AWS General Reference/ for the following:

For a list of service name strings, see AWS Service Namespaces.

For resource type strings, see Example ARNs.

For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

You can specify multiple resource types by using a comma separated array. The array can include up to 100 items. Note that the length constraint requirement applies to each resource type filter.

tagKeyFilters :: Maybe (NonEmpty Text)

Specifies that you want the response to include information for only resources that have tags with the specified tag keys. If you use this parameter, the count of returned noncompliant resources includes only resources that have the specified tag keys.

newGetComplianceSummary :: GetComplianceSummary

Create a value of GetComplianceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxResults:GetComplianceSummary', getComplianceSummary_maxResults - Specifies the maximum number of results to be returned in each page. A query can return fewer than this maximum, even if there are more results still to return. You should always check the PaginationToken response value to see if there are more results. You can specify a minimum of 1 and a maximum value of 100.

$sel:regionFilters:GetComplianceSummary', getComplianceSummary_regionFilters - Specifies a list of AWS Regions to limit the output by. If you use this parameter, the count of returned noncompliant resources includes only resources in the specified Regions.

$sel:paginationToken:GetComplianceSummary', getComplianceSummary_paginationToken - Specifies a PaginationToken response value from a previous request to indicate that you want the next page of results. Leave this parameter empty in your initial request.

$sel:targetIdFilters:GetComplianceSummary', getComplianceSummary_targetIdFilters - Specifies target identifiers (usually, specific account IDs) to limit the output by. If you use this parameter, the count of returned noncompliant resources includes only resources with the specified target IDs.

$sel:groupBy:GetComplianceSummary', getComplianceSummary_groupBy - Specifies a list of attributes to group the counts of noncompliant resources by. If supplied, the counts are sorted by those attributes.

$sel:resourceTypeFilters:GetComplianceSummary', getComplianceSummary_resourceTypeFilters - Specifies that you want the response to include information for only resources of the specified types. The format of each resource type is service[:resourceType]. For example, specifying a resource type of ec2 returns all Amazon EC2 resources (which includes EC2 instances). Specifying a resource type of ec2:instance returns only EC2 instances.

The string for each service name and resource type is the same as that embedded in a resource's Amazon Resource Name (ARN). Consult the /AWS General Reference/ for the following:

For a list of service name strings, see AWS Service Namespaces.

For resource type strings, see Example ARNs.

For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

You can specify multiple resource types by using a comma separated array. The array can include up to 100 items. Note that the length constraint requirement applies to each resource type filter.

$sel:tagKeyFilters:GetComplianceSummary', getComplianceSummary_tagKeyFilters - Specifies that you want the response to include information for only resources that have tags with the specified tag keys. If you use this parameter, the count of returned noncompliant resources includes only resources that have the specified tag keys.

Request Lenses

getComplianceSummary_maxResults :: Lens' GetComplianceSummary (Maybe Natural)

Specifies the maximum number of results to be returned in each page. A query can return fewer than this maximum, even if there are more results still to return. You should always check the PaginationToken response value to see if there are more results. You can specify a minimum of 1 and a maximum value of 100.

getComplianceSummary_regionFilters :: Lens' GetComplianceSummary (Maybe (NonEmpty Text))

Specifies a list of AWS Regions to limit the output by. If you use this parameter, the count of returned noncompliant resources includes only resources in the specified Regions.

getComplianceSummary_paginationToken :: Lens' GetComplianceSummary (Maybe Text)

Specifies a PaginationToken response value from a previous request to indicate that you want the next page of results. Leave this parameter empty in your initial request.

getComplianceSummary_targetIdFilters :: Lens' GetComplianceSummary (Maybe (NonEmpty Text))

Specifies target identifiers (usually, specific account IDs) to limit the output by. If you use this parameter, the count of returned noncompliant resources includes only resources with the specified target IDs.

getComplianceSummary_groupBy :: Lens' GetComplianceSummary (Maybe [GroupByAttribute])

Specifies a list of attributes to group the counts of noncompliant resources by. If supplied, the counts are sorted by those attributes.

getComplianceSummary_resourceTypeFilters :: Lens' GetComplianceSummary (Maybe [Text])

Specifies that you want the response to include information for only resources of the specified types. The format of each resource type is service[:resourceType]. For example, specifying a resource type of ec2 returns all Amazon EC2 resources (which includes EC2 instances). Specifying a resource type of ec2:instance returns only EC2 instances.

The string for each service name and resource type is the same as that embedded in a resource's Amazon Resource Name (ARN). Consult the /AWS General Reference/ for the following:

For a list of service name strings, see AWS Service Namespaces.

For resource type strings, see Example ARNs.

For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

You can specify multiple resource types by using a comma separated array. The array can include up to 100 items. Note that the length constraint requirement applies to each resource type filter.

getComplianceSummary_tagKeyFilters :: Lens' GetComplianceSummary (Maybe (NonEmpty Text))

Specifies that you want the response to include information for only resources that have tags with the specified tag keys. If you use this parameter, the count of returned noncompliant resources includes only resources that have the specified tag keys.

Destructuring the Response

data GetComplianceSummaryResponse

See: newGetComplianceSummaryResponse smart constructor.

Constructors

GetComplianceSummaryResponse'

Fields

paginationToken :: Maybe Text

A string that indicates that there is more data available than this response contains. To receive the next part of the response, specify this response value as the PaginationToken value in the request for the next page.

summaryList :: Maybe [Summary]

A table that shows counts of noncompliant resources.

httpStatus :: Int

The response's http status code.

newGetComplianceSummaryResponse

Arguments

:: Int
$sel:httpStatus:GetComplianceSummaryResponse'

-> GetComplianceSummaryResponse

Create a value of GetComplianceSummaryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:paginationToken:GetComplianceSummary', getComplianceSummaryResponse_paginationToken - A string that indicates that there is more data available than this response contains. To receive the next part of the response, specify this response value as the PaginationToken value in the request for the next page.

$sel:summaryList:GetComplianceSummaryResponse', getComplianceSummaryResponse_summaryList - A table that shows counts of noncompliant resources.

$sel:httpStatus:GetComplianceSummaryResponse', getComplianceSummaryResponse_httpStatus - The response's http status code.

Response Lenses

getComplianceSummaryResponse_paginationToken :: Lens' GetComplianceSummaryResponse (Maybe Text)

A string that indicates that there is more data available than this response contains. To receive the next part of the response, specify this response value as the PaginationToken value in the request for the next page.

getComplianceSummaryResponse_summaryList :: Lens' GetComplianceSummaryResponse (Maybe [Summary])

A table that shows counts of noncompliant resources.

getComplianceSummaryResponse_httpStatus :: Lens' GetComplianceSummaryResponse Int

The response's http status code.