Network.AWS.ResourceGroupsTagging.GetResources

Description

Returns all the tagged or previously tagged resources that are located in the specified Region for the AWS account.

Depending on what information you want returned, you can also specify the following:

Filters that specify what tags and resource types you want returned. The response includes all tags that are associated with the requested resources.

Information about compliance with the account's effective tag policy. For more information on tag policies, see Tag Policies in the AWS Organizations User Guide.

This operation supports pagination, where the response can be sent in multiple pages. You should check the PaginationToken response parameter to determine if there are additional results available to return. Repeat the query, passing the PaginationToken response parameter value as an input to the next request until you recieve a null value. A null value for PaginationToken indicates that there are no more results waiting to be returned.

This operation returns paginated results.

Creating a Request

data GetResources

See: newGetResources smart constructor.

Constructors

GetResources'

Fields

paginationToken :: Maybe Text

Specifies a PaginationToken response value from a previous request to indicate that you want the next page of results. Leave this parameter empty in your initial request.

resourcesPerPage :: Maybe Int

Specifies the maximum number of results to be returned in each page. A query can return fewer than this maximum, even if there are more results still to return. You should always check the PaginationToken response value to see if there are more results. You can specify a minimum of 1 and a maximum value of 100.

excludeCompliantResources :: Maybe Bool

Specifies whether to exclude resources that are compliant with the tag policy. Set this to true if you are interested in retrieving information on noncompliant resources only.

You can use this parameter only if the IncludeComplianceDetails parameter is also set to true.

includeComplianceDetails :: Maybe Bool

Specifies whether to include details regarding the compliance with the effective tag policy. Set this to true to determine whether resources are compliant with the tag policy and to get details.

resourceTypeFilters :: Maybe [Text]

Specifies the resource types that you want included in the response. The format of each resource type is service[:resourceType]. For example, specifying a resource type of ec2 returns all Amazon EC2 resources (which includes EC2 instances). Specifying a resource type of ec2:instance returns only EC2 instances.

The string for each service name and resource type is the same as that embedded in a resource's Amazon Resource Name (ARN). Consult the /AWS General Reference/ for the following:

For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

You can specify multiple resource types by using an array. The array can include up to 100 items. Note that the length constraint requirement applies to each resource type filter.

tagFilters :: Maybe [TagFilter]

Specifies a list of TagFilters (keys and values) to restrict the output to only those resources that have the specified tag and, if included, the specified value. Each TagFilter must contain a key with values optional. A request can include up to 50 keys, and each key can include up to 20 values.

Note the following when deciding how to use TagFilters:

If you don't specify a TagFilter, the response includes all resources that are currently tagged or ever had a tag. Resources that currently don't have tags are shown with an empty tag set, like this: "Tags": [].

If you specify more than one filter in a single request, the response returns only those resources that satisfy all filters.

If you specify a filter that contains more than one value for a key, the response returns resources that match any of the specified values for that key.

If you don't specify any values for a key, the response returns resources that are tagged with that key and any or no value.

For example, for the following filters: filter1= {keyA,{value1}}, filter2={keyB,{value2,value3,value4}}, filter3= {keyC}:

GetResources({filter1}) returns resources tagged with key1=value1

GetResources({filter2}) returns resources tagged with key2=value2 or key2=value3 or key2=value4

GetResources({filter3}) returns resources tagged with any tag with the key key3, and with any or no value

GetResources({filter1,filter2,filter3}) returns resources tagged with (key1=value1) and (key2=value2 or key2=value3 or key2=value4) and (key3, any or no value)

resourceARNList :: Maybe (NonEmpty Text)

Specifies a list of ARNs of resources for which you want to retrieve tag data. You can't specify both this parameter and any of the pagination parameters (ResourcesPerPage, TagsPerPage, PaginationToken) in the same request. If you specify both, you get an Invalid Parameter exception.

If a resource specified by this parameter doesn't exist, it doesn't generate an error; it simply isn't included in the response.

An ARN (Amazon Resource Name) uniquely identifies a resource. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

tagsPerPage :: Maybe Int

AWS recommends using ResourcesPerPage instead of this parameter.

A limit that restricts the number of tags (key and value pairs) returned by GetResources in paginated output. A resource with no tags is counted as having one tag (one key and value pair).

GetResources does not split a resource and its associated tags across pages. If the specified TagsPerPage would cause such a break, a PaginationToken is returned in place of the affected resource and its tags. Use that token in another request to get the remaining data. For example, if you specify a TagsPerPage of 100 and the account has 22 resources with 10 tags each (meaning that each resource has 10 key and value pairs), the output will consist of three pages. The first page displays the first 10 resources, each with its 10 tags. The second page displays the next 10 resources, each with its 10 tags. The third page displays the remaining 2 resources, each with its 10 tags.

You can set TagsPerPage to a minimum of 100 items up to a maximum of 500 items.

newGetResources :: GetResources

Create a value of GetResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:paginationToken:GetResources', getResources_paginationToken - Specifies a PaginationToken response value from a previous request to indicate that you want the next page of results. Leave this parameter empty in your initial request.

$sel:resourcesPerPage:GetResources', getResources_resourcesPerPage - Specifies the maximum number of results to be returned in each page. A query can return fewer than this maximum, even if there are more results still to return. You should always check the PaginationToken response value to see if there are more results. You can specify a minimum of 1 and a maximum value of 100.

$sel:excludeCompliantResources:GetResources', getResources_excludeCompliantResources - Specifies whether to exclude resources that are compliant with the tag policy. Set this to true if you are interested in retrieving information on noncompliant resources only.

You can use this parameter only if the IncludeComplianceDetails parameter is also set to true.

$sel:includeComplianceDetails:GetResources', getResources_includeComplianceDetails - Specifies whether to include details regarding the compliance with the effective tag policy. Set this to true to determine whether resources are compliant with the tag policy and to get details.

$sel:resourceTypeFilters:GetResources', getResources_resourceTypeFilters - Specifies the resource types that you want included in the response. The format of each resource type is service[:resourceType]. For example, specifying a resource type of ec2 returns all Amazon EC2 resources (which includes EC2 instances). Specifying a resource type of ec2:instance returns only EC2 instances.

The string for each service name and resource type is the same as that embedded in a resource's Amazon Resource Name (ARN). Consult the /AWS General Reference/ for the following:

For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

You can specify multiple resource types by using an array. The array can include up to 100 items. Note that the length constraint requirement applies to each resource type filter.

$sel:tagFilters:GetResources', getResources_tagFilters - Specifies a list of TagFilters (keys and values) to restrict the output to only those resources that have the specified tag and, if included, the specified value. Each TagFilter must contain a key with values optional. A request can include up to 50 keys, and each key can include up to 20 values.

Note the following when deciding how to use TagFilters:

If you don't specify a TagFilter, the response includes all resources that are currently tagged or ever had a tag. Resources that currently don't have tags are shown with an empty tag set, like this: "Tags": [].

If you specify more than one filter in a single request, the response returns only those resources that satisfy all filters.

If you specify a filter that contains more than one value for a key, the response returns resources that match any of the specified values for that key.

If you don't specify any values for a key, the response returns resources that are tagged with that key and any or no value.

For example, for the following filters: filter1= {keyA,{value1}}, filter2={keyB,{value2,value3,value4}}, filter3= {keyC}:

GetResources({filter1}) returns resources tagged with key1=value1

GetResources({filter2}) returns resources tagged with key2=value2 or key2=value3 or key2=value4

GetResources({filter3}) returns resources tagged with any tag with the key key3, and with any or no value

GetResources({filter1,filter2,filter3}) returns resources tagged with (key1=value1) and (key2=value2 or key2=value3 or key2=value4) and (key3, any or no value)

$sel:resourceARNList:GetResources', getResources_resourceARNList - Specifies a list of ARNs of resources for which you want to retrieve tag data. You can't specify both this parameter and any of the pagination parameters (ResourcesPerPage, TagsPerPage, PaginationToken) in the same request. If you specify both, you get an Invalid Parameter exception.

If a resource specified by this parameter doesn't exist, it doesn't generate an error; it simply isn't included in the response.

An ARN (Amazon Resource Name) uniquely identifies a resource. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

$sel:tagsPerPage:GetResources', getResources_tagsPerPage - AWS recommends using ResourcesPerPage instead of this parameter.

A limit that restricts the number of tags (key and value pairs) returned by GetResources in paginated output. A resource with no tags is counted as having one tag (one key and value pair).

GetResources does not split a resource and its associated tags across pages. If the specified TagsPerPage would cause such a break, a PaginationToken is returned in place of the affected resource and its tags. Use that token in another request to get the remaining data. For example, if you specify a TagsPerPage of 100 and the account has 22 resources with 10 tags each (meaning that each resource has 10 key and value pairs), the output will consist of three pages. The first page displays the first 10 resources, each with its 10 tags. The second page displays the next 10 resources, each with its 10 tags. The third page displays the remaining 2 resources, each with its 10 tags.

You can set TagsPerPage to a minimum of 100 items up to a maximum of 500 items.

Request Lenses

getResources_paginationToken :: Lens' GetResources (Maybe Text)

Specifies a PaginationToken response value from a previous request to indicate that you want the next page of results. Leave this parameter empty in your initial request.

getResources_resourcesPerPage :: Lens' GetResources (Maybe Int)

Specifies the maximum number of results to be returned in each page. A query can return fewer than this maximum, even if there are more results still to return. You should always check the PaginationToken response value to see if there are more results. You can specify a minimum of 1 and a maximum value of 100.

getResources_excludeCompliantResources :: Lens' GetResources (Maybe Bool)

Specifies whether to exclude resources that are compliant with the tag policy. Set this to true if you are interested in retrieving information on noncompliant resources only.

You can use this parameter only if the IncludeComplianceDetails parameter is also set to true.

getResources_includeComplianceDetails :: Lens' GetResources (Maybe Bool)

Specifies whether to include details regarding the compliance with the effective tag policy. Set this to true to determine whether resources are compliant with the tag policy and to get details.

getResources_resourceTypeFilters :: Lens' GetResources (Maybe [Text])

Specifies the resource types that you want included in the response. The format of each resource type is service[:resourceType]. For example, specifying a resource type of ec2 returns all Amazon EC2 resources (which includes EC2 instances). Specifying a resource type of ec2:instance returns only EC2 instances.

The string for each service name and resource type is the same as that embedded in a resource's Amazon Resource Name (ARN). Consult the /AWS General Reference/ for the following:

For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

You can specify multiple resource types by using an array. The array can include up to 100 items. Note that the length constraint requirement applies to each resource type filter.

getResources_tagFilters :: Lens' GetResources (Maybe [TagFilter])

Specifies a list of TagFilters (keys and values) to restrict the output to only those resources that have the specified tag and, if included, the specified value. Each TagFilter must contain a key with values optional. A request can include up to 50 keys, and each key can include up to 20 values.

Note the following when deciding how to use TagFilters:

If you don't specify a TagFilter, the response includes all resources that are currently tagged or ever had a tag. Resources that currently don't have tags are shown with an empty tag set, like this: "Tags": [].

If you specify more than one filter in a single request, the response returns only those resources that satisfy all filters.

If you specify a filter that contains more than one value for a key, the response returns resources that match any of the specified values for that key.

If you don't specify any values for a key, the response returns resources that are tagged with that key and any or no value.

For example, for the following filters: filter1= {keyA,{value1}}, filter2={keyB,{value2,value3,value4}}, filter3= {keyC}:

GetResources({filter1}) returns resources tagged with key1=value1

GetResources({filter2}) returns resources tagged with key2=value2 or key2=value3 or key2=value4

GetResources({filter3}) returns resources tagged with any tag with the key key3, and with any or no value

GetResources({filter1,filter2,filter3}) returns resources tagged with (key1=value1) and (key2=value2 or key2=value3 or key2=value4) and (key3, any or no value)

getResources_resourceARNList :: Lens' GetResources (Maybe (NonEmpty Text))

Specifies a list of ARNs of resources for which you want to retrieve tag data. You can't specify both this parameter and any of the pagination parameters (ResourcesPerPage, TagsPerPage, PaginationToken) in the same request. If you specify both, you get an Invalid Parameter exception.

If a resource specified by this parameter doesn't exist, it doesn't generate an error; it simply isn't included in the response.

An ARN (Amazon Resource Name) uniquely identifies a resource. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

getResources_tagsPerPage :: Lens' GetResources (Maybe Int)

AWS recommends using ResourcesPerPage instead of this parameter.

A limit that restricts the number of tags (key and value pairs) returned by GetResources in paginated output. A resource with no tags is counted as having one tag (one key and value pair).

GetResources does not split a resource and its associated tags across pages. If the specified TagsPerPage would cause such a break, a PaginationToken is returned in place of the affected resource and its tags. Use that token in another request to get the remaining data. For example, if you specify a TagsPerPage of 100 and the account has 22 resources with 10 tags each (meaning that each resource has 10 key and value pairs), the output will consist of three pages. The first page displays the first 10 resources, each with its 10 tags. The second page displays the next 10 resources, each with its 10 tags. The third page displays the remaining 2 resources, each with its 10 tags.

You can set TagsPerPage to a minimum of 100 items up to a maximum of 500 items.

Destructuring the Response

data GetResourcesResponse

See: newGetResourcesResponse smart constructor.

Constructors

GetResourcesResponse'

Fields

paginationToken :: Maybe Text

A string that indicates that there is more data available than this response contains. To receive the next part of the response, specify this response value as the PaginationToken value in the request for the next page.

resourceTagMappingList :: Maybe [ResourceTagMapping]

A list of resource ARNs and the tags (keys and values) associated with those ARNs.

httpStatus :: Int

The response's http status code.

newGetResourcesResponse

Arguments

:: Int
$sel:httpStatus:GetResourcesResponse'

-> GetResourcesResponse

Create a value of GetResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:paginationToken:GetResources', getResourcesResponse_paginationToken - A string that indicates that there is more data available than this response contains. To receive the next part of the response, specify this response value as the PaginationToken value in the request for the next page.

$sel:resourceTagMappingList:GetResourcesResponse', getResourcesResponse_resourceTagMappingList - A list of resource ARNs and the tags (keys and values) associated with those ARNs.

$sel:httpStatus:GetResourcesResponse', getResourcesResponse_httpStatus - The response's http status code.

Response Lenses

getResourcesResponse_paginationToken :: Lens' GetResourcesResponse (Maybe Text)

A string that indicates that there is more data available than this response contains. To receive the next part of the response, specify this response value as the PaginationToken value in the request for the next page.

getResourcesResponse_resourceTagMappingList :: Lens' GetResourcesResponse (Maybe [ResourceTagMapping])

A list of resource ARNs and the tags (keys and values) associated with those ARNs.

getResourcesResponse_httpStatus :: Lens' GetResourcesResponse Int

The response's http status code.