Network.AWS.ResourceGroupsTagging.Lens

Description

Operations

GetComplianceSummary

getComplianceSummary_groupBy :: Lens' GetComplianceSummary (Maybe [GroupByAttribute])

Specifies a list of attributes to group the counts of noncompliant resources by. If supplied, the counts are sorted by those attributes.

getComplianceSummary_paginationToken :: Lens' GetComplianceSummary (Maybe Text)

Specifies a PaginationToken response value from a previous request to indicate that you want the next page of results. Leave this parameter empty in your initial request.

getComplianceSummary_targetIdFilters :: Lens' GetComplianceSummary (Maybe (NonEmpty Text))

Specifies target identifiers (usually, specific account IDs) to limit the output by. If you use this parameter, the count of returned noncompliant resources includes only resources with the specified target IDs.

getComplianceSummary_resourceTypeFilters :: Lens' GetComplianceSummary (Maybe [Text])

Specifies that you want the response to include information for only resources of the specified types. The format of each resource type is service[:resourceType]. For example, specifying a resource type of ec2 returns all Amazon EC2 resources (which includes EC2 instances). Specifying a resource type of ec2:instance returns only EC2 instances.

The string for each service name and resource type is the same as that embedded in a resource's Amazon Resource Name (ARN). Consult the /AWS General Reference/ for the following:

For a list of service name strings, see AWS Service Namespaces.

For resource type strings, see Example ARNs.

For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

You can specify multiple resource types by using a comma separated array. The array can include up to 100 items. Note that the length constraint requirement applies to each resource type filter.

getComplianceSummary_regionFilters :: Lens' GetComplianceSummary (Maybe (NonEmpty Text))

Specifies a list of AWS Regions to limit the output by. If you use this parameter, the count of returned noncompliant resources includes only resources in the specified Regions.

getComplianceSummary_tagKeyFilters :: Lens' GetComplianceSummary (Maybe (NonEmpty Text))

Specifies that you want the response to include information for only resources that have tags with the specified tag keys. If you use this parameter, the count of returned noncompliant resources includes only resources that have the specified tag keys.

getComplianceSummary_maxResults :: Lens' GetComplianceSummary (Maybe Natural)

Specifies the maximum number of results to be returned in each page. A query can return fewer than this maximum, even if there are more results still to return. You should always check the PaginationToken response value to see if there are more results. You can specify a minimum of 1 and a maximum value of 100.

getComplianceSummaryResponse_paginationToken :: Lens' GetComplianceSummaryResponse (Maybe Text)

A string that indicates that there is more data available than this response contains. To receive the next part of the response, specify this response value as the PaginationToken value in the request for the next page.

getComplianceSummaryResponse_summaryList :: Lens' GetComplianceSummaryResponse (Maybe [Summary])

A table that shows counts of noncompliant resources.

getComplianceSummaryResponse_httpStatus :: Lens' GetComplianceSummaryResponse Int

The response's http status code.

DescribeReportCreation

describeReportCreationResponse_status :: Lens' DescribeReportCreationResponse (Maybe Text)

Reports the status of the operation.

The operation status can be one of the following:

RUNNING - Report creation is in progress.

SUCCEEDED - Report creation is complete. You can open the report from the Amazon S3 bucket that you specified when you ran StartReportCreation.

FAILED - Report creation timed out or the Amazon S3 bucket is not accessible.

NO REPORT - No report was generated in the last 90 days.

describeReportCreationResponse_s3Location :: Lens' DescribeReportCreationResponse (Maybe Text)

The path to the Amazon S3 bucket where the report was stored on creation.

describeReportCreationResponse_errorMessage :: Lens' DescribeReportCreationResponse (Maybe Text)

Details of the common errors that all operations return.

describeReportCreationResponse_httpStatus :: Lens' DescribeReportCreationResponse Int

The response's http status code.

GetTagKeys

getTagKeys_paginationToken :: Lens' GetTagKeys (Maybe Text)

Specifies a PaginationToken response value from a previous request to indicate that you want the next page of results. Leave this parameter empty in your initial request.

getTagKeysResponse_paginationToken :: Lens' GetTagKeysResponse (Maybe Text)

A string that indicates that there is more data available than this response contains. To receive the next part of the response, specify this response value as the PaginationToken value in the request for the next page.

getTagKeysResponse_tagKeys :: Lens' GetTagKeysResponse (Maybe [Text])

A list of all tag keys in the AWS account.

getTagKeysResponse_httpStatus :: Lens' GetTagKeysResponse Int

The response's http status code.

TagResources

tagResources_resourceARNList :: Lens' TagResources (NonEmpty Text)

Specifies the list of ARNs of the resources that you want to apply tags to.

An ARN (Amazon Resource Name) uniquely identifies a resource. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

tagResources_tags :: Lens' TagResources (HashMap Text Text)

Specifies a list of tags that you want to add to the specified resources. A tag consists of a key and a value that you define.

tagResourcesResponse_failedResourcesMap :: Lens' TagResourcesResponse (Maybe (HashMap Text FailureInfo))

A map containing a key-value pair for each failed item that couldn't be tagged. The key is the ARN of the failed resource. The value is a FailureInfo object that contains an error code, a status code, and an error message. If there are no errors, the FailedResourcesMap is empty.

tagResourcesResponse_httpStatus :: Lens' TagResourcesResponse Int

The response's http status code.

GetTagValues

getTagValues_paginationToken :: Lens' GetTagValues (Maybe Text)

Specifies a PaginationToken response value from a previous request to indicate that you want the next page of results. Leave this parameter empty in your initial request.

getTagValues_key :: Lens' GetTagValues Text

Specifies the tag key for which you want to list all existing values that are currently used in the specified AWS Region for the calling AWS account.

getTagValuesResponse_paginationToken :: Lens' GetTagValuesResponse (Maybe Text)

A string that indicates that there is more data available than this response contains. To receive the next part of the response, specify this response value as the PaginationToken value in the request for the next page.

getTagValuesResponse_tagValues :: Lens' GetTagValuesResponse (Maybe [Text])

A list of all tag values for the specified key currently used in the specified AWS Region for the calling AWS account.

getTagValuesResponse_httpStatus :: Lens' GetTagValuesResponse Int

The response's http status code.

StartReportCreation

startReportCreation_s3Bucket :: Lens' StartReportCreation Text

The name of the Amazon S3 bucket where the report will be stored; for example:

awsexamplebucket

For more information on S3 bucket requirements, including an example bucket policy, see the example S3 bucket policy on this page.

startReportCreationResponse_httpStatus :: Lens' StartReportCreationResponse Int

The response's http status code.

GetResources

getResources_paginationToken :: Lens' GetResources (Maybe Text)

Specifies a PaginationToken response value from a previous request to indicate that you want the next page of results. Leave this parameter empty in your initial request.

getResources_resourcesPerPage :: Lens' GetResources (Maybe Int)

Specifies the maximum number of results to be returned in each page. A query can return fewer than this maximum, even if there are more results still to return. You should always check the PaginationToken response value to see if there are more results. You can specify a minimum of 1 and a maximum value of 100.

getResources_excludeCompliantResources :: Lens' GetResources (Maybe Bool)

Specifies whether to exclude resources that are compliant with the tag policy. Set this to true if you are interested in retrieving information on noncompliant resources only.

You can use this parameter only if the IncludeComplianceDetails parameter is also set to true.

getResources_includeComplianceDetails :: Lens' GetResources (Maybe Bool)

Specifies whether to include details regarding the compliance with the effective tag policy. Set this to true to determine whether resources are compliant with the tag policy and to get details.

getResources_resourceTypeFilters :: Lens' GetResources (Maybe [Text])

Specifies the resource types that you want included in the response. The format of each resource type is service[:resourceType]. For example, specifying a resource type of ec2 returns all Amazon EC2 resources (which includes EC2 instances). Specifying a resource type of ec2:instance returns only EC2 instances.

The string for each service name and resource type is the same as that embedded in a resource's Amazon Resource Name (ARN). Consult the /AWS General Reference/ for the following:

For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

You can specify multiple resource types by using an array. The array can include up to 100 items. Note that the length constraint requirement applies to each resource type filter.

getResources_tagFilters :: Lens' GetResources (Maybe [TagFilter])

Specifies a list of TagFilters (keys and values) to restrict the output to only those resources that have the specified tag and, if included, the specified value. Each TagFilter must contain a key with values optional. A request can include up to 50 keys, and each key can include up to 20 values.

Note the following when deciding how to use TagFilters:

If you don't specify a TagFilter, the response includes all resources that are currently tagged or ever had a tag. Resources that currently don't have tags are shown with an empty tag set, like this: "Tags": [].

If you specify more than one filter in a single request, the response returns only those resources that satisfy all filters.

If you specify a filter that contains more than one value for a key, the response returns resources that match any of the specified values for that key.

If you don't specify any values for a key, the response returns resources that are tagged with that key and any or no value.

For example, for the following filters: filter1= {keyA,{value1}}, filter2={keyB,{value2,value3,value4}}, filter3= {keyC}:

GetResources({filter1}) returns resources tagged with key1=value1

GetResources({filter2}) returns resources tagged with key2=value2 or key2=value3 or key2=value4

GetResources({filter3}) returns resources tagged with any tag with the key key3, and with any or no value

GetResources({filter1,filter2,filter3}) returns resources tagged with (key1=value1) and (key2=value2 or key2=value3 or key2=value4) and (key3, any or no value)

getResources_resourceARNList :: Lens' GetResources (Maybe (NonEmpty Text))

Specifies a list of ARNs of resources for which you want to retrieve tag data. You can't specify both this parameter and any of the pagination parameters (ResourcesPerPage, TagsPerPage, PaginationToken) in the same request. If you specify both, you get an Invalid Parameter exception.

If a resource specified by this parameter doesn't exist, it doesn't generate an error; it simply isn't included in the response.

An ARN (Amazon Resource Name) uniquely identifies a resource. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

getResources_tagsPerPage :: Lens' GetResources (Maybe Int)

AWS recommends using ResourcesPerPage instead of this parameter.

A limit that restricts the number of tags (key and value pairs) returned by GetResources in paginated output. A resource with no tags is counted as having one tag (one key and value pair).

GetResources does not split a resource and its associated tags across pages. If the specified TagsPerPage would cause such a break, a PaginationToken is returned in place of the affected resource and its tags. Use that token in another request to get the remaining data. For example, if you specify a TagsPerPage of 100 and the account has 22 resources with 10 tags each (meaning that each resource has 10 key and value pairs), the output will consist of three pages. The first page displays the first 10 resources, each with its 10 tags. The second page displays the next 10 resources, each with its 10 tags. The third page displays the remaining 2 resources, each with its 10 tags.

You can set TagsPerPage to a minimum of 100 items up to a maximum of 500 items.

getResourcesResponse_paginationToken :: Lens' GetResourcesResponse (Maybe Text)

A string that indicates that there is more data available than this response contains. To receive the next part of the response, specify this response value as the PaginationToken value in the request for the next page.

getResourcesResponse_resourceTagMappingList :: Lens' GetResourcesResponse (Maybe [ResourceTagMapping])

A list of resource ARNs and the tags (keys and values) associated with those ARNs.

getResourcesResponse_httpStatus :: Lens' GetResourcesResponse Int

The response's http status code.

UntagResources

untagResources_resourceARNList :: Lens' UntagResources (NonEmpty Text)

Specifies a list of ARNs of the resources that you want to remove tags from.

An ARN (Amazon Resource Name) uniquely identifies a resource. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

untagResources_tagKeys :: Lens' UntagResources (NonEmpty Text)

Specifies a list of tag keys that you want to remove from the specified resources.

untagResourcesResponse_failedResourcesMap :: Lens' UntagResourcesResponse (Maybe (HashMap Text FailureInfo))

A map containing a key-value pair for each failed item that couldn't be untagged. The key is the ARN of the failed resource. The value is a FailureInfo object that contains an error code, a status code, and an error message. If there are no errors, the FailedResourcesMap is empty.

untagResourcesResponse_httpStatus :: Lens' UntagResourcesResponse Int

The response's http status code.

Types

ComplianceDetails

complianceDetails_keysWithNoncompliantValues :: Lens' ComplianceDetails (Maybe [Text])

These are keys defined in the effective policy that are on the resource with either incorrect case treatment or noncompliant values.

complianceDetails_complianceStatus :: Lens' ComplianceDetails (Maybe Bool)

Whether a resource is compliant with the effective tag policy.

complianceDetails_noncompliantKeys :: Lens' ComplianceDetails (Maybe [Text])

These tag keys on the resource are noncompliant with the effective tag policy.

FailureInfo

failureInfo_errorCode :: Lens' FailureInfo (Maybe ResourceErrorCode)

The code of the common error. Valid values include InternalServiceException, InvalidParameterException, and any valid error code returned by the AWS service that hosts the resource that you want to tag.

failureInfo_errorMessage :: Lens' FailureInfo (Maybe Text)

The message of the common error.

failureInfo_statusCode :: Lens' FailureInfo (Maybe Int)

The HTTP status code of the common error.

ResourceTagMapping

resourceTagMapping_complianceDetails :: Lens' ResourceTagMapping (Maybe ComplianceDetails)

Information that shows whether a resource is compliant with the effective tag policy, including details on any noncompliant tag keys.

resourceTagMapping_resourceARN :: Lens' ResourceTagMapping (Maybe Text)

The ARN of the resource.

resourceTagMapping_tags :: Lens' ResourceTagMapping (Maybe [Tag])

The tags that have been applied to one or more AWS resources.

Summary

summary_targetId :: Lens' Summary (Maybe Text)

The account identifier or the root identifier of the organization. If you don't know the root ID, you can call the AWS Organizations ListRoots API.

summary_lastUpdated :: Lens' Summary (Maybe Text)

The timestamp that shows when this summary was generated in this Region.

summary_resourceType :: Lens' Summary (Maybe Text)

The AWS resource type.

summary_nonCompliantResources :: Lens' Summary (Maybe Integer)

The count of noncompliant resources.

summary_targetIdType :: Lens' Summary (Maybe TargetIdType)

Whether the target is an account, an OU, or the organization root.

summary_region :: Lens' Summary (Maybe Text)

The AWS Region that the summary applies to.

Tag

tag_key :: Lens' Tag Text

One part of a key-value pair that makes up a tag. A key is a general label that acts like a category for more specific tag values.

tag_value :: Lens' Tag Text

One part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key). The value can be empty or null.

TagFilter

tagFilter_values :: Lens' TagFilter (Maybe [Text])

One part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key). The value can be empty or null.

tagFilter_key :: Lens' TagFilter (Maybe Text)

One part of a key-value pair that makes up a tag. A key is a general label that acts like a category for more specific tag values.