Network.AWS.ResourceGroupsTagging.Types

Description

Service Configuration

defaultService :: Service

API version 2017-01-26 of the Amazon Resource Groups Tagging API SDK configuration.

Errors

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

This error indicates one of the following:

A parameter is missing.

A malformed string was supplied for the request parameter.

An out-of-range value was supplied for the request parameter.

The target ID is invalid, unsupported, or doesn't exist.

You can't access the Amazon S3 bucket for report storage. For more information, see Additional Requirements for Organization-wide Tag Compliance Reports in the AWS Organizations User Guide.

_ConstraintViolationException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied because performing this operation violates a constraint.

Some of the reasons in the following list might not apply to this specific operation.

You must meet the prerequisites for using tag policies. For information, see Prerequisites and Permissions for Using Tag Policies in the AWS Organizations User Guide.

You must enable the tag policies service principal (tagpolicies.tag.amazonaws.com) to integrate with AWS Organizations For information, see EnableAWSServiceAccess.

You must have a tag policy attached to the organization root, an OU, or an account.

_ThrottledException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied to limit the frequency of submitted requests.

_PaginationTokenExpiredException :: AsError a => Getting (First ServiceError) a ServiceError

A PaginationToken is valid for a maximum of 15 minutes. Your request was denied because the specified PaginationToken has expired.

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

The target of the operation is currently being modified by a different request. Try again later.

_InternalServiceException :: AsError a => Getting (First ServiceError) a ServiceError

The request processing failed because of an unknown error, exception, or failure. You can retry the request.

GroupByAttribute

newtype GroupByAttribute

Constructors

GroupByAttribute'

Fields

fromGroupByAttribute :: Text

Bundled Patterns

pattern GroupByAttribute_REGION :: GroupByAttribute

pattern GroupByAttribute_RESOURCE_TYPE :: GroupByAttribute

pattern GroupByAttribute_TARGET_ID :: GroupByAttribute

ResourceErrorCode

newtype ResourceErrorCode

Constructors

ResourceErrorCode'

Fields

fromResourceErrorCode :: Text

Bundled Patterns

pattern ResourceErrorCode_InternalServiceException :: ResourceErrorCode

pattern ResourceErrorCode_InvalidParameterException :: ResourceErrorCode

TargetIdType

newtype TargetIdType

Constructors

TargetIdType'

Fields

fromTargetIdType :: Text

Bundled Patterns

pattern TargetIdType_ACCOUNT :: TargetIdType

pattern TargetIdType_OU :: TargetIdType

pattern TargetIdType_ROOT :: TargetIdType

ComplianceDetails

data ComplianceDetails

Information that shows whether a resource is compliant with the effective tag policy, including details on any noncompliant tag keys.

See: newComplianceDetails smart constructor.

Constructors

ComplianceDetails'

Fields

keysWithNoncompliantValues :: Maybe [Text]

These are keys defined in the effective policy that are on the resource with either incorrect case treatment or noncompliant values.

complianceStatus :: Maybe Bool

Whether a resource is compliant with the effective tag policy.

noncompliantKeys :: Maybe [Text]

These tag keys on the resource are noncompliant with the effective tag policy.

newComplianceDetails :: ComplianceDetails

Create a value of ComplianceDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keysWithNoncompliantValues:ComplianceDetails', complianceDetails_keysWithNoncompliantValues - These are keys defined in the effective policy that are on the resource with either incorrect case treatment or noncompliant values.

$sel:complianceStatus:ComplianceDetails', complianceDetails_complianceStatus - Whether a resource is compliant with the effective tag policy.

$sel:noncompliantKeys:ComplianceDetails', complianceDetails_noncompliantKeys - These tag keys on the resource are noncompliant with the effective tag policy.

complianceDetails_keysWithNoncompliantValues :: Lens' ComplianceDetails (Maybe [Text])

These are keys defined in the effective policy that are on the resource with either incorrect case treatment or noncompliant values.

complianceDetails_complianceStatus :: Lens' ComplianceDetails (Maybe Bool)

Whether a resource is compliant with the effective tag policy.

complianceDetails_noncompliantKeys :: Lens' ComplianceDetails (Maybe [Text])

These tag keys on the resource are noncompliant with the effective tag policy.

FailureInfo

data FailureInfo

Information about the errors that are returned for each failed resource. This information can include InternalServiceException and InvalidParameterException errors. It can also include any valid error code returned by the AWS service that hosts the resource that the ARN key represents.

The following are common error codes that you might receive from other AWS services:

InternalServiceException – This can mean that the Resource Groups Tagging API didn't receive a response from another AWS service. It can also mean the the resource type in the request is not supported by the Resource Groups Tagging API. In these cases, it's safe to retry the request and then call GetResources to verify the changes.

AccessDeniedException – This can mean that you need permission to calling tagging operations in the AWS service that contains the resource. For example, to use the Resource Groups Tagging API to tag a CloudWatch alarm resource, you need permission to call TagResources and TagResource in the CloudWatch API.

For more information on errors that are generated from other AWS services, see the documentation for that service.

See: newFailureInfo smart constructor.

Constructors

FailureInfo'

Fields

errorCode :: Maybe ResourceErrorCode

The code of the common error. Valid values include InternalServiceException, InvalidParameterException, and any valid error code returned by the AWS service that hosts the resource that you want to tag.

errorMessage :: Maybe Text

The message of the common error.

statusCode :: Maybe Int

The HTTP status code of the common error.

newFailureInfo :: FailureInfo

Create a value of FailureInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCode:FailureInfo', failureInfo_errorCode - The code of the common error. Valid values include InternalServiceException, InvalidParameterException, and any valid error code returned by the AWS service that hosts the resource that you want to tag.

$sel:errorMessage:FailureInfo', failureInfo_errorMessage - The message of the common error.

$sel:statusCode:FailureInfo', failureInfo_statusCode - The HTTP status code of the common error.

failureInfo_errorCode :: Lens' FailureInfo (Maybe ResourceErrorCode)

The code of the common error. Valid values include InternalServiceException, InvalidParameterException, and any valid error code returned by the AWS service that hosts the resource that you want to tag.

failureInfo_errorMessage :: Lens' FailureInfo (Maybe Text)

The message of the common error.

failureInfo_statusCode :: Lens' FailureInfo (Maybe Int)

The HTTP status code of the common error.

ResourceTagMapping

data ResourceTagMapping

A list of resource ARNs and the tags (keys and values) that are associated with each.

See: newResourceTagMapping smart constructor.

Constructors

ResourceTagMapping'

Fields

complianceDetails :: Maybe ComplianceDetails

Information that shows whether a resource is compliant with the effective tag policy, including details on any noncompliant tag keys.

resourceARN :: Maybe Text

The ARN of the resource.

tags :: Maybe [Tag]

The tags that have been applied to one or more AWS resources.

newResourceTagMapping :: ResourceTagMapping

Create a value of ResourceTagMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:complianceDetails:ResourceTagMapping', resourceTagMapping_complianceDetails - Information that shows whether a resource is compliant with the effective tag policy, including details on any noncompliant tag keys.

$sel:resourceARN:ResourceTagMapping', resourceTagMapping_resourceARN - The ARN of the resource.

$sel:tags:ResourceTagMapping', resourceTagMapping_tags - The tags that have been applied to one or more AWS resources.

resourceTagMapping_complianceDetails :: Lens' ResourceTagMapping (Maybe ComplianceDetails)

Information that shows whether a resource is compliant with the effective tag policy, including details on any noncompliant tag keys.

resourceTagMapping_resourceARN :: Lens' ResourceTagMapping (Maybe Text)

The ARN of the resource.

resourceTagMapping_tags :: Lens' ResourceTagMapping (Maybe [Tag])

The tags that have been applied to one or more AWS resources.

Summary

data Summary

A count of noncompliant resources.

See: newSummary smart constructor.

Constructors

Summary'

Fields

targetId :: Maybe Text

The account identifier or the root identifier of the organization. If you don't know the root ID, you can call the AWS Organizations ListRoots API.

lastUpdated :: Maybe Text

The timestamp that shows when this summary was generated in this Region.

resourceType :: Maybe Text

The AWS resource type.

nonCompliantResources :: Maybe Integer

The count of noncompliant resources.

targetIdType :: Maybe TargetIdType

Whether the target is an account, an OU, or the organization root.

region :: Maybe Text

The AWS Region that the summary applies to.

newSummary :: Summary

Create a value of Summary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetId:Summary', summary_targetId - The account identifier or the root identifier of the organization. If you don't know the root ID, you can call the AWS Organizations ListRoots API.

$sel:lastUpdated:Summary', summary_lastUpdated - The timestamp that shows when this summary was generated in this Region.

$sel:resourceType:Summary', summary_resourceType - The AWS resource type.

$sel:nonCompliantResources:Summary', summary_nonCompliantResources - The count of noncompliant resources.

$sel:targetIdType:Summary', summary_targetIdType - Whether the target is an account, an OU, or the organization root.

$sel:region:Summary', summary_region - The AWS Region that the summary applies to.

summary_targetId :: Lens' Summary (Maybe Text)

The account identifier or the root identifier of the organization. If you don't know the root ID, you can call the AWS Organizations ListRoots API.

summary_lastUpdated :: Lens' Summary (Maybe Text)

The timestamp that shows when this summary was generated in this Region.

summary_resourceType :: Lens' Summary (Maybe Text)

The AWS resource type.

summary_nonCompliantResources :: Lens' Summary (Maybe Integer)

The count of noncompliant resources.

summary_targetIdType :: Lens' Summary (Maybe TargetIdType)

Whether the target is an account, an OU, or the organization root.

summary_region :: Lens' Summary (Maybe Text)

The AWS Region that the summary applies to.

Tag

data Tag

The metadata that you apply to AWS resources to help you categorize and organize them. Each tag consists of a key and a value, both of which you define. For more information, see Tagging AWS Resources in the AWS General Reference.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

One part of a key-value pair that makes up a tag. A key is a general label that acts like a category for more specific tag values.

value :: Text

One part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key). The value can be empty or null.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - One part of a key-value pair that makes up a tag. A key is a general label that acts like a category for more specific tag values.

$sel:value:Tag', tag_value - One part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key). The value can be empty or null.

tag_key :: Lens' Tag Text

One part of a key-value pair that makes up a tag. A key is a general label that acts like a category for more specific tag values.

tag_value :: Lens' Tag Text

One part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key). The value can be empty or null.

TagFilter

data TagFilter

A list of tags (keys and values) that are used to specify the associated resources.

See: newTagFilter smart constructor.

Constructors

TagFilter'

Fields

values :: Maybe [Text]

One part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key). The value can be empty or null.

key :: Maybe Text

One part of a key-value pair that makes up a tag. A key is a general label that acts like a category for more specific tag values.

newTagFilter :: TagFilter

Create a value of TagFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:TagFilter', tagFilter_values - One part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key). The value can be empty or null.

$sel:key:TagFilter', tagFilter_key - One part of a key-value pair that makes up a tag. A key is a general label that acts like a category for more specific tag values.

tagFilter_values :: Lens' TagFilter (Maybe [Text])

One part of a key-value pair that make up a tag. A value acts as a descriptor within a tag category (key). The value can be empty or null.

tagFilter_key :: Lens' TagFilter (Maybe Text)

One part of a key-value pair that makes up a tag. A key is a general label that acts like a category for more specific tag values.