Network.AWS.ResourceGroupsTagging.Types.FailureInfo

Description

Documentation

data FailureInfo

Information about the errors that are returned for each failed resource. This information can include InternalServiceException and InvalidParameterException errors. It can also include any valid error code returned by the AWS service that hosts the resource that the ARN key represents.

The following are common error codes that you might receive from other AWS services:

InternalServiceException – This can mean that the Resource Groups Tagging API didn't receive a response from another AWS service. It can also mean the the resource type in the request is not supported by the Resource Groups Tagging API. In these cases, it's safe to retry the request and then call GetResources to verify the changes.

AccessDeniedException – This can mean that you need permission to calling tagging operations in the AWS service that contains the resource. For example, to use the Resource Groups Tagging API to tag a CloudWatch alarm resource, you need permission to call TagResources and TagResource in the CloudWatch API.

For more information on errors that are generated from other AWS services, see the documentation for that service.

See: newFailureInfo smart constructor.

Constructors

FailureInfo'

Fields

statusCode :: Maybe Int

The HTTP status code of the common error.

errorMessage :: Maybe Text

The message of the common error.

errorCode :: Maybe ResourceErrorCode

The code of the common error. Valid values include InternalServiceException, InvalidParameterException, and any valid error code returned by the AWS service that hosts the resource that you want to tag.

newFailureInfo :: FailureInfo

Create a value of FailureInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusCode:FailureInfo', failureInfo_statusCode - The HTTP status code of the common error.

$sel:errorMessage:FailureInfo', failureInfo_errorMessage - The message of the common error.

$sel:errorCode:FailureInfo', failureInfo_errorCode - The code of the common error. Valid values include InternalServiceException, InvalidParameterException, and any valid error code returned by the AWS service that hosts the resource that you want to tag.

failureInfo_statusCode :: Lens' FailureInfo (Maybe Int)

The HTTP status code of the common error.

failureInfo_errorMessage :: Lens' FailureInfo (Maybe Text)

The message of the common error.

failureInfo_errorCode :: Lens' FailureInfo (Maybe ResourceErrorCode)

The code of the common error. Valid values include InternalServiceException, InvalidParameterException, and any valid error code returned by the AWS service that hosts the resource that you want to tag.