Network.AWS.ResourceGroupsTagging.UntagResources

Description

Removes the specified tags from the specified resources. When you specify a tag key, the action removes both that key and its associated value. The operation succeeds even if you attempt to remove tags from a resource that were already removed. Note the following:

To remove tags from a resource, you need the necessary permissions for the service that the resource belongs to as well as permissions for removing tags. For more information, see the documentation for the service whose resource you want to untag.

You can only tag resources that are located in the specified AWS Region for the calling AWS account.

Creating a Request

data UntagResources

See: newUntagResources smart constructor.

Constructors

UntagResources'

Fields

resourceARNList :: NonEmpty Text

Specifies a list of ARNs of the resources that you want to remove tags from.

An ARN (Amazon Resource Name) uniquely identifies a resource. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

tagKeys :: NonEmpty Text

Specifies a list of tag keys that you want to remove from the specified resources.

newUntagResources

Arguments

:: NonEmpty Text
$sel:resourceARNList:UntagResources'

-> NonEmpty Text
$sel:tagKeys:UntagResources'

-> UntagResources

Create a value of UntagResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceARNList:UntagResources', untagResources_resourceARNList - Specifies a list of ARNs of the resources that you want to remove tags from.

An ARN (Amazon Resource Name) uniquely identifies a resource. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

$sel:tagKeys:UntagResources', untagResources_tagKeys - Specifies a list of tag keys that you want to remove from the specified resources.

Request Lenses

untagResources_resourceARNList :: Lens' UntagResources (NonEmpty Text)

Specifies a list of ARNs of the resources that you want to remove tags from.

An ARN (Amazon Resource Name) uniquely identifies a resource. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

untagResources_tagKeys :: Lens' UntagResources (NonEmpty Text)

Specifies a list of tag keys that you want to remove from the specified resources.

Destructuring the Response

data UntagResourcesResponse

See: newUntagResourcesResponse smart constructor.

Constructors

UntagResourcesResponse'

Fields

failedResourcesMap :: Maybe (HashMap Text FailureInfo)

A map containing a key-value pair for each failed item that couldn't be untagged. The key is the ARN of the failed resource. The value is a FailureInfo object that contains an error code, a status code, and an error message. If there are no errors, the FailedResourcesMap is empty.

httpStatus :: Int

The response's http status code.

newUntagResourcesResponse

Arguments

:: Int
$sel:httpStatus:UntagResourcesResponse'

-> UntagResourcesResponse

Create a value of UntagResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedResourcesMap:UntagResourcesResponse', untagResourcesResponse_failedResourcesMap - A map containing a key-value pair for each failed item that couldn't be untagged. The key is the ARN of the failed resource. The value is a FailureInfo object that contains an error code, a status code, and an error message. If there are no errors, the FailedResourcesMap is empty.

$sel:httpStatus:UntagResourcesResponse', untagResourcesResponse_httpStatus - The response's http status code.

Response Lenses

untagResourcesResponse_failedResourcesMap :: Lens' UntagResourcesResponse (Maybe (HashMap Text FailureInfo))

A map containing a key-value pair for each failed item that couldn't be untagged. The key is the ARN of the failed resource. The value is a FailureInfo object that contains an error code, a status code, and an error message. If there are no errors, the FailedResourcesMap is empty.

untagResourcesResponse_httpStatus :: Lens' UntagResourcesResponse Int

The response's http status code.