Network.AWS.RobOMaker.CreateRobot

Description

Creates a robot.

Creating a Request

data CreateRobot

See: newCreateRobot smart constructor.

Constructors

CreateRobot'

Fields

tags :: Maybe (HashMap Text Text)

A map that contains tag keys and tag values that are attached to the robot.

name :: Text

The name for the robot.

architecture :: Architecture

The target architecture of the robot.

greengrassGroupId :: Text

The Greengrass group id.

newCreateRobot

Arguments

:: Text
$sel:name:CreateRobot'

-> Architecture
$sel:architecture:CreateRobot'

-> Text
$sel:greengrassGroupId:CreateRobot'

-> CreateRobot

Create a value of CreateRobot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateRobot', createRobot_tags - A map that contains tag keys and tag values that are attached to the robot.

$sel:name:CreateRobot', createRobot_name - The name for the robot.

$sel:architecture:CreateRobot', createRobot_architecture - The target architecture of the robot.

$sel:greengrassGroupId:CreateRobot', createRobot_greengrassGroupId - The Greengrass group id.

Request Lenses

createRobot_tags :: Lens' CreateRobot (Maybe (HashMap Text Text))

A map that contains tag keys and tag values that are attached to the robot.

createRobot_name :: Lens' CreateRobot Text

The name for the robot.

createRobot_architecture :: Lens' CreateRobot Architecture

The target architecture of the robot.

createRobot_greengrassGroupId :: Lens' CreateRobot Text

The Greengrass group id.

Destructuring the Response

data CreateRobotResponse

See: newCreateRobotResponse smart constructor.

Constructors

CreateRobotResponse'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the robot.

createdAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the robot was created.

greengrassGroupId :: Maybe Text

The Amazon Resource Name (ARN) of the Greengrass group associated with the robot.

name :: Maybe Text

The name of the robot.

architecture :: Maybe Architecture

The target architecture of the robot.

tags :: Maybe (HashMap Text Text)

The list of all tags added to the robot.

httpStatus :: Int

The response's http status code.

newCreateRobotResponse

Arguments

:: Int
$sel:httpStatus:CreateRobotResponse'

-> CreateRobotResponse

Create a value of CreateRobotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateRobotResponse', createRobotResponse_arn - The Amazon Resource Name (ARN) of the robot.

$sel:createdAt:CreateRobotResponse', createRobotResponse_createdAt - The time, in milliseconds since the epoch, when the robot was created.

$sel:greengrassGroupId:CreateRobot', createRobotResponse_greengrassGroupId - The Amazon Resource Name (ARN) of the Greengrass group associated with the robot.

$sel:name:CreateRobot', createRobotResponse_name - The name of the robot.

$sel:architecture:CreateRobot', createRobotResponse_architecture - The target architecture of the robot.

$sel:tags:CreateRobot', createRobotResponse_tags - The list of all tags added to the robot.

$sel:httpStatus:CreateRobotResponse', createRobotResponse_httpStatus - The response's http status code.

Response Lenses

createRobotResponse_arn :: Lens' CreateRobotResponse (Maybe Text)

The Amazon Resource Name (ARN) of the robot.

createRobotResponse_createdAt :: Lens' CreateRobotResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the robot was created.

createRobotResponse_greengrassGroupId :: Lens' CreateRobotResponse (Maybe Text)

The Amazon Resource Name (ARN) of the Greengrass group associated with the robot.

createRobotResponse_name :: Lens' CreateRobotResponse (Maybe Text)

The name of the robot.

createRobotResponse_architecture :: Lens' CreateRobotResponse (Maybe Architecture)

The target architecture of the robot.

createRobotResponse_tags :: Lens' CreateRobotResponse (Maybe (HashMap Text Text))

The list of all tags added to the robot.

createRobotResponse_httpStatus :: Lens' CreateRobotResponse Int

The response's http status code.