Network.AWS.RobOMaker.CreateRobotApplicationVersion

Description

Creates a version of a robot application.

Creating a Request

data CreateRobotApplicationVersion

See: newCreateRobotApplicationVersion smart constructor.

Constructors

CreateRobotApplicationVersion'

Fields

currentRevisionId :: Maybe Text

The current revision id for the robot application. If you provide a value and it matches the latest revision ID, a new version will be created.

s3Etags :: Maybe [Text]

The Amazon S3 identifier for the zip file bundle that you use for your robot application.

imageDigest :: Maybe Text

A SHA256 identifier for the Docker image that you use for your robot application.

application :: Text

The application information for the robot application.

newCreateRobotApplicationVersion

Arguments

:: Text
$sel:application:CreateRobotApplicationVersion'

-> CreateRobotApplicationVersion

Create a value of CreateRobotApplicationVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currentRevisionId:CreateRobotApplicationVersion', createRobotApplicationVersion_currentRevisionId - The current revision id for the robot application. If you provide a value and it matches the latest revision ID, a new version will be created.

$sel:s3Etags:CreateRobotApplicationVersion', createRobotApplicationVersion_s3Etags - The Amazon S3 identifier for the zip file bundle that you use for your robot application.

$sel:imageDigest:CreateRobotApplicationVersion', createRobotApplicationVersion_imageDigest - A SHA256 identifier for the Docker image that you use for your robot application.

$sel:application:CreateRobotApplicationVersion', createRobotApplicationVersion_application - The application information for the robot application.

Request Lenses

createRobotApplicationVersion_currentRevisionId :: Lens' CreateRobotApplicationVersion (Maybe Text)

The current revision id for the robot application. If you provide a value and it matches the latest revision ID, a new version will be created.

createRobotApplicationVersion_s3Etags :: Lens' CreateRobotApplicationVersion (Maybe [Text])

The Amazon S3 identifier for the zip file bundle that you use for your robot application.

createRobotApplicationVersion_imageDigest :: Lens' CreateRobotApplicationVersion (Maybe Text)

A SHA256 identifier for the Docker image that you use for your robot application.

createRobotApplicationVersion_application :: Lens' CreateRobotApplicationVersion Text

The application information for the robot application.

Destructuring the Response

data CreateRobotApplicationVersionResponse

See: newCreateRobotApplicationVersionResponse smart constructor.

Constructors

CreateRobotApplicationVersionResponse'

Fields

lastUpdatedAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the robot application was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the robot application.

environment :: Maybe Environment

The object that contains the Docker image URI used to create your robot application.

sources :: Maybe [Source]

The sources of the robot application.

name :: Maybe Text

The name of the robot application.

version :: Maybe Text

The version of the robot application.

robotSoftwareSuite :: Maybe RobotSoftwareSuite

The robot software suite (ROS distribution) used by the robot application.

revisionId :: Maybe Text

The revision id of the robot application.

httpStatus :: Int

The response's http status code.

newCreateRobotApplicationVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateRobotApplicationVersionResponse'

-> CreateRobotApplicationVersionResponse

Create a value of CreateRobotApplicationVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedAt:CreateRobotApplicationVersionResponse', createRobotApplicationVersionResponse_lastUpdatedAt - The time, in milliseconds since the epoch, when the robot application was last updated.

$sel:arn:CreateRobotApplicationVersionResponse', createRobotApplicationVersionResponse_arn - The Amazon Resource Name (ARN) of the robot application.

$sel:environment:CreateRobotApplicationVersionResponse', createRobotApplicationVersionResponse_environment - The object that contains the Docker image URI used to create your robot application.

$sel:sources:CreateRobotApplicationVersionResponse', createRobotApplicationVersionResponse_sources - The sources of the robot application.

$sel:name:CreateRobotApplicationVersionResponse', createRobotApplicationVersionResponse_name - The name of the robot application.

$sel:version:CreateRobotApplicationVersionResponse', createRobotApplicationVersionResponse_version - The version of the robot application.

$sel:robotSoftwareSuite:CreateRobotApplicationVersionResponse', createRobotApplicationVersionResponse_robotSoftwareSuite - The robot software suite (ROS distribution) used by the robot application.

$sel:revisionId:CreateRobotApplicationVersionResponse', createRobotApplicationVersionResponse_revisionId - The revision id of the robot application.

$sel:httpStatus:CreateRobotApplicationVersionResponse', createRobotApplicationVersionResponse_httpStatus - The response's http status code.

Response Lenses

createRobotApplicationVersionResponse_lastUpdatedAt :: Lens' CreateRobotApplicationVersionResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the robot application was last updated.

createRobotApplicationVersionResponse_arn :: Lens' CreateRobotApplicationVersionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the robot application.

createRobotApplicationVersionResponse_environment :: Lens' CreateRobotApplicationVersionResponse (Maybe Environment)

The object that contains the Docker image URI used to create your robot application.

createRobotApplicationVersionResponse_sources :: Lens' CreateRobotApplicationVersionResponse (Maybe [Source])

The sources of the robot application.

createRobotApplicationVersionResponse_name :: Lens' CreateRobotApplicationVersionResponse (Maybe Text)

The name of the robot application.

createRobotApplicationVersionResponse_version :: Lens' CreateRobotApplicationVersionResponse (Maybe Text)

The version of the robot application.

createRobotApplicationVersionResponse_robotSoftwareSuite :: Lens' CreateRobotApplicationVersionResponse (Maybe RobotSoftwareSuite)

The robot software suite (ROS distribution) used by the robot application.

createRobotApplicationVersionResponse_revisionId :: Lens' CreateRobotApplicationVersionResponse (Maybe Text)

The revision id of the robot application.

createRobotApplicationVersionResponse_httpStatus :: Lens' CreateRobotApplicationVersionResponse Int

The response's http status code.