Network.AWS.RobOMaker.CreateSimulationApplicationVersion

Description

Creates a simulation application with a specific revision id.

Creating a Request

data CreateSimulationApplicationVersion

See: newCreateSimulationApplicationVersion smart constructor.

Constructors

CreateSimulationApplicationVersion'

Fields

currentRevisionId :: Maybe Text

The current revision id for the simulation application. If you provide a value and it matches the latest revision ID, a new version will be created.

s3Etags :: Maybe [Text]

The Amazon S3 eTag identifier for the zip file bundle that you use to create the simulation application.

imageDigest :: Maybe Text

The SHA256 digest used to identify the Docker image URI used to created the simulation application.

application :: Text

The application information for the simulation application.

newCreateSimulationApplicationVersion

Arguments

:: Text
$sel:application:CreateSimulationApplicationVersion'

-> CreateSimulationApplicationVersion

Create a value of CreateSimulationApplicationVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currentRevisionId:CreateSimulationApplicationVersion', createSimulationApplicationVersion_currentRevisionId - The current revision id for the simulation application. If you provide a value and it matches the latest revision ID, a new version will be created.

$sel:s3Etags:CreateSimulationApplicationVersion', createSimulationApplicationVersion_s3Etags - The Amazon S3 eTag identifier for the zip file bundle that you use to create the simulation application.

$sel:imageDigest:CreateSimulationApplicationVersion', createSimulationApplicationVersion_imageDigest - The SHA256 digest used to identify the Docker image URI used to created the simulation application.

$sel:application:CreateSimulationApplicationVersion', createSimulationApplicationVersion_application - The application information for the simulation application.

Request Lenses

createSimulationApplicationVersion_currentRevisionId :: Lens' CreateSimulationApplicationVersion (Maybe Text)

The current revision id for the simulation application. If you provide a value and it matches the latest revision ID, a new version will be created.

createSimulationApplicationVersion_s3Etags :: Lens' CreateSimulationApplicationVersion (Maybe [Text])

The Amazon S3 eTag identifier for the zip file bundle that you use to create the simulation application.

createSimulationApplicationVersion_imageDigest :: Lens' CreateSimulationApplicationVersion (Maybe Text)

The SHA256 digest used to identify the Docker image URI used to created the simulation application.

createSimulationApplicationVersion_application :: Lens' CreateSimulationApplicationVersion Text

The application information for the simulation application.

Destructuring the Response

data CreateSimulationApplicationVersionResponse

See: newCreateSimulationApplicationVersionResponse smart constructor.

Constructors

CreateSimulationApplicationVersionResponse'

Fields

renderingEngine :: Maybe RenderingEngine

The rendering engine for the simulation application.

lastUpdatedAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the simulation application was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the simulation application.

environment :: Maybe Environment

The object that contains the Docker image URI used to create the simulation application.

sources :: Maybe [Source]

The sources of the simulation application.

name :: Maybe Text

The name of the simulation application.

version :: Maybe Text

The version of the simulation application.

simulationSoftwareSuite :: Maybe SimulationSoftwareSuite

The simulation software suite used by the simulation application.

robotSoftwareSuite :: Maybe RobotSoftwareSuite

Information about the robot software suite (ROS distribution).

revisionId :: Maybe Text

The revision ID of the simulation application.

httpStatus :: Int

The response's http status code.

newCreateSimulationApplicationVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateSimulationApplicationVersionResponse'

-> CreateSimulationApplicationVersionResponse

Create a value of CreateSimulationApplicationVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:renderingEngine:CreateSimulationApplicationVersionResponse', createSimulationApplicationVersionResponse_renderingEngine - The rendering engine for the simulation application.

$sel:lastUpdatedAt:CreateSimulationApplicationVersionResponse', createSimulationApplicationVersionResponse_lastUpdatedAt - The time, in milliseconds since the epoch, when the simulation application was last updated.

$sel:arn:CreateSimulationApplicationVersionResponse', createSimulationApplicationVersionResponse_arn - The Amazon Resource Name (ARN) of the simulation application.

$sel:environment:CreateSimulationApplicationVersionResponse', createSimulationApplicationVersionResponse_environment - The object that contains the Docker image URI used to create the simulation application.

$sel:sources:CreateSimulationApplicationVersionResponse', createSimulationApplicationVersionResponse_sources - The sources of the simulation application.

$sel:name:CreateSimulationApplicationVersionResponse', createSimulationApplicationVersionResponse_name - The name of the simulation application.

$sel:version:CreateSimulationApplicationVersionResponse', createSimulationApplicationVersionResponse_version - The version of the simulation application.

$sel:simulationSoftwareSuite:CreateSimulationApplicationVersionResponse', createSimulationApplicationVersionResponse_simulationSoftwareSuite - The simulation software suite used by the simulation application.

$sel:robotSoftwareSuite:CreateSimulationApplicationVersionResponse', createSimulationApplicationVersionResponse_robotSoftwareSuite - Information about the robot software suite (ROS distribution).

$sel:revisionId:CreateSimulationApplicationVersionResponse', createSimulationApplicationVersionResponse_revisionId - The revision ID of the simulation application.

$sel:httpStatus:CreateSimulationApplicationVersionResponse', createSimulationApplicationVersionResponse_httpStatus - The response's http status code.

Response Lenses

createSimulationApplicationVersionResponse_renderingEngine :: Lens' CreateSimulationApplicationVersionResponse (Maybe RenderingEngine)

The rendering engine for the simulation application.

createSimulationApplicationVersionResponse_lastUpdatedAt :: Lens' CreateSimulationApplicationVersionResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the simulation application was last updated.

createSimulationApplicationVersionResponse_arn :: Lens' CreateSimulationApplicationVersionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the simulation application.

createSimulationApplicationVersionResponse_environment :: Lens' CreateSimulationApplicationVersionResponse (Maybe Environment)

The object that contains the Docker image URI used to create the simulation application.

createSimulationApplicationVersionResponse_sources :: Lens' CreateSimulationApplicationVersionResponse (Maybe [Source])

The sources of the simulation application.

createSimulationApplicationVersionResponse_name :: Lens' CreateSimulationApplicationVersionResponse (Maybe Text)

The name of the simulation application.

createSimulationApplicationVersionResponse_version :: Lens' CreateSimulationApplicationVersionResponse (Maybe Text)

The version of the simulation application.

createSimulationApplicationVersionResponse_simulationSoftwareSuite :: Lens' CreateSimulationApplicationVersionResponse (Maybe SimulationSoftwareSuite)

The simulation software suite used by the simulation application.

createSimulationApplicationVersionResponse_robotSoftwareSuite :: Lens' CreateSimulationApplicationVersionResponse (Maybe RobotSoftwareSuite)

Information about the robot software suite (ROS distribution).

createSimulationApplicationVersionResponse_revisionId :: Lens' CreateSimulationApplicationVersionResponse (Maybe Text)

The revision ID of the simulation application.

createSimulationApplicationVersionResponse_httpStatus :: Lens' CreateSimulationApplicationVersionResponse Int

The response's http status code.