Network.AWS.RobOMaker.CreateWorldGenerationJob

Description

Creates worlds using the specified template.

Creating a Request

data CreateWorldGenerationJob

See: newCreateWorldGenerationJob smart constructor.

Constructors

CreateWorldGenerationJob'

Fields

worldTags :: Maybe (HashMap Text Text)

A map that contains tag keys and tag values that are attached to the generated worlds.

clientRequestToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

tags :: Maybe (HashMap Text Text)

A map that contains tag keys and tag values that are attached to the world generator job.

template :: Text

The Amazon Resource Name (arn) of the world template describing the worlds you want to create.

worldCount :: WorldCount

Information about the world count.

newCreateWorldGenerationJob

Arguments

:: Text
$sel:template:CreateWorldGenerationJob'

-> WorldCount
$sel:worldCount:CreateWorldGenerationJob'

-> CreateWorldGenerationJob

Create a value of CreateWorldGenerationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:worldTags:CreateWorldGenerationJob', createWorldGenerationJob_worldTags - A map that contains tag keys and tag values that are attached to the generated worlds.

$sel:clientRequestToken:CreateWorldGenerationJob', createWorldGenerationJob_clientRequestToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:tags:CreateWorldGenerationJob', createWorldGenerationJob_tags - A map that contains tag keys and tag values that are attached to the world generator job.

$sel:template:CreateWorldGenerationJob', createWorldGenerationJob_template - The Amazon Resource Name (arn) of the world template describing the worlds you want to create.

$sel:worldCount:CreateWorldGenerationJob', createWorldGenerationJob_worldCount - Information about the world count.

Request Lenses

createWorldGenerationJob_worldTags :: Lens' CreateWorldGenerationJob (Maybe (HashMap Text Text))

A map that contains tag keys and tag values that are attached to the generated worlds.

createWorldGenerationJob_clientRequestToken :: Lens' CreateWorldGenerationJob (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createWorldGenerationJob_tags :: Lens' CreateWorldGenerationJob (Maybe (HashMap Text Text))

A map that contains tag keys and tag values that are attached to the world generator job.

createWorldGenerationJob_template :: Lens' CreateWorldGenerationJob Text

The Amazon Resource Name (arn) of the world template describing the worlds you want to create.

createWorldGenerationJob_worldCount :: Lens' CreateWorldGenerationJob WorldCount

Information about the world count.

Destructuring the Response

data CreateWorldGenerationJobResponse

See: newCreateWorldGenerationJobResponse smart constructor.

Constructors

CreateWorldGenerationJobResponse'

Fields

status :: Maybe WorldGenerationJobStatus

The status of the world generator job.

Pending

The world generator job request is pending.

Running

The world generator job is running.

Completed

The world generator job completed.

Failed

The world generator job failed. See failureCode for more information.

PartialFailed

Some worlds did not generate.

Canceled

The world generator job was cancelled.

Canceling

The world generator job is being cancelled.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the world generator job.

createdAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the world generator job was created.

failureCode :: Maybe WorldGenerationJobErrorCode

The failure code of the world generator job if it failed:

InternalServiceError

Internal service error.

LimitExceeded

The requested resource exceeds the maximum number allowed, or the number of concurrent stream requests exceeds the maximum number allowed.

ResourceNotFound

The specified resource could not be found.

RequestThrottled

The request was throttled.

InvalidInput

An input parameter in the request is not valid.

worldCount :: Maybe WorldCount

Information about the world count.

template :: Maybe Text

The Amazon Resource Name (arn) of the world template.

worldTags :: Maybe (HashMap Text Text)

A map that contains tag keys and tag values that are attached to the generated worlds.

clientRequestToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

tags :: Maybe (HashMap Text Text)

A map that contains tag keys and tag values that are attached to the world generator job.

httpStatus :: Int

The response's http status code.

newCreateWorldGenerationJobResponse

Arguments

:: Int
$sel:httpStatus:CreateWorldGenerationJobResponse'

-> CreateWorldGenerationJobResponse

Create a value of CreateWorldGenerationJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateWorldGenerationJobResponse', createWorldGenerationJobResponse_status - The status of the world generator job.

Pending

The world generator job request is pending.

Running

The world generator job is running.

Completed

The world generator job completed.

Failed

The world generator job failed. See failureCode for more information.

PartialFailed

Some worlds did not generate.

Canceled

The world generator job was cancelled.

Canceling

The world generator job is being cancelled.

$sel:arn:CreateWorldGenerationJobResponse', createWorldGenerationJobResponse_arn - The Amazon Resource Name (ARN) of the world generator job.

$sel:createdAt:CreateWorldGenerationJobResponse', createWorldGenerationJobResponse_createdAt - The time, in milliseconds since the epoch, when the world generator job was created.

$sel:failureCode:CreateWorldGenerationJobResponse', createWorldGenerationJobResponse_failureCode - The failure code of the world generator job if it failed:

InternalServiceError

Internal service error.

LimitExceeded

The requested resource exceeds the maximum number allowed, or the number of concurrent stream requests exceeds the maximum number allowed.

ResourceNotFound

The specified resource could not be found.

RequestThrottled

The request was throttled.

InvalidInput

An input parameter in the request is not valid.

$sel:worldCount:CreateWorldGenerationJob', createWorldGenerationJobResponse_worldCount - Information about the world count.

$sel:template:CreateWorldGenerationJob', createWorldGenerationJobResponse_template - The Amazon Resource Name (arn) of the world template.

$sel:worldTags:CreateWorldGenerationJob', createWorldGenerationJobResponse_worldTags - A map that contains tag keys and tag values that are attached to the generated worlds.

$sel:clientRequestToken:CreateWorldGenerationJob', createWorldGenerationJobResponse_clientRequestToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:tags:CreateWorldGenerationJob', createWorldGenerationJobResponse_tags - A map that contains tag keys and tag values that are attached to the world generator job.

$sel:httpStatus:CreateWorldGenerationJobResponse', createWorldGenerationJobResponse_httpStatus - The response's http status code.

Response Lenses

createWorldGenerationJobResponse_status :: Lens' CreateWorldGenerationJobResponse (Maybe WorldGenerationJobStatus)

The status of the world generator job.

Pending

The world generator job request is pending.

Running

The world generator job is running.

Completed

The world generator job completed.

Failed

The world generator job failed. See failureCode for more information.

PartialFailed

Some worlds did not generate.

Canceled

The world generator job was cancelled.

Canceling

The world generator job is being cancelled.

createWorldGenerationJobResponse_arn :: Lens' CreateWorldGenerationJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the world generator job.

createWorldGenerationJobResponse_createdAt :: Lens' CreateWorldGenerationJobResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the world generator job was created.

createWorldGenerationJobResponse_failureCode :: Lens' CreateWorldGenerationJobResponse (Maybe WorldGenerationJobErrorCode)

The failure code of the world generator job if it failed:

InternalServiceError

Internal service error.

LimitExceeded

The requested resource exceeds the maximum number allowed, or the number of concurrent stream requests exceeds the maximum number allowed.

ResourceNotFound

The specified resource could not be found.

RequestThrottled

The request was throttled.

InvalidInput

An input parameter in the request is not valid.

createWorldGenerationJobResponse_worldCount :: Lens' CreateWorldGenerationJobResponse (Maybe WorldCount)

Information about the world count.

createWorldGenerationJobResponse_template :: Lens' CreateWorldGenerationJobResponse (Maybe Text)

The Amazon Resource Name (arn) of the world template.

createWorldGenerationJobResponse_worldTags :: Lens' CreateWorldGenerationJobResponse (Maybe (HashMap Text Text))

A map that contains tag keys and tag values that are attached to the generated worlds.

createWorldGenerationJobResponse_clientRequestToken :: Lens' CreateWorldGenerationJobResponse (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createWorldGenerationJobResponse_tags :: Lens' CreateWorldGenerationJobResponse (Maybe (HashMap Text Text))

A map that contains tag keys and tag values that are attached to the world generator job.

createWorldGenerationJobResponse_httpStatus :: Lens' CreateWorldGenerationJobResponse Int

The response's http status code.