Network.AWS.RobOMaker.DescribeRobotApplication

Description

Describes a robot application.

Creating a Request

data DescribeRobotApplication

See: newDescribeRobotApplication smart constructor.

Constructors

DescribeRobotApplication'

Fields

applicationVersion :: Maybe Text

The version of the robot application to describe.

application :: Text

The Amazon Resource Name (ARN) of the robot application.

newDescribeRobotApplication

Arguments

:: Text
$sel:application:DescribeRobotApplication'

-> DescribeRobotApplication

Create a value of DescribeRobotApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationVersion:DescribeRobotApplication', describeRobotApplication_applicationVersion - The version of the robot application to describe.

$sel:application:DescribeRobotApplication', describeRobotApplication_application - The Amazon Resource Name (ARN) of the robot application.

Request Lenses

describeRobotApplication_applicationVersion :: Lens' DescribeRobotApplication (Maybe Text)

The version of the robot application to describe.

describeRobotApplication_application :: Lens' DescribeRobotApplication Text

The Amazon Resource Name (ARN) of the robot application.

Destructuring the Response

data DescribeRobotApplicationResponse

See: newDescribeRobotApplicationResponse smart constructor.

Constructors

DescribeRobotApplicationResponse'

Fields

lastUpdatedAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the robot application was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the robot application.

environment :: Maybe Environment

The object that contains the Docker image URI used to create the robot application.

sources :: Maybe [Source]

The sources of the robot application.

name :: Maybe Text

The name of the robot application.

imageDigest :: Maybe Text

A SHA256 identifier for the Docker image that you use for your robot application.

version :: Maybe Text

The version of the robot application.

robotSoftwareSuite :: Maybe RobotSoftwareSuite

The robot software suite (ROS distribution) used by the robot application.

revisionId :: Maybe Text

The revision id of the robot application.

tags :: Maybe (HashMap Text Text)

The list of all tags added to the specified robot application.

httpStatus :: Int

The response's http status code.

newDescribeRobotApplicationResponse

Arguments

:: Int
$sel:httpStatus:DescribeRobotApplicationResponse'

-> DescribeRobotApplicationResponse

Create a value of DescribeRobotApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedAt:DescribeRobotApplicationResponse', describeRobotApplicationResponse_lastUpdatedAt - The time, in milliseconds since the epoch, when the robot application was last updated.

$sel:arn:DescribeRobotApplicationResponse', describeRobotApplicationResponse_arn - The Amazon Resource Name (ARN) of the robot application.

$sel:environment:DescribeRobotApplicationResponse', describeRobotApplicationResponse_environment - The object that contains the Docker image URI used to create the robot application.

$sel:sources:DescribeRobotApplicationResponse', describeRobotApplicationResponse_sources - The sources of the robot application.

$sel:name:DescribeRobotApplicationResponse', describeRobotApplicationResponse_name - The name of the robot application.

$sel:imageDigest:DescribeRobotApplicationResponse', describeRobotApplicationResponse_imageDigest - A SHA256 identifier for the Docker image that you use for your robot application.

$sel:version:DescribeRobotApplicationResponse', describeRobotApplicationResponse_version - The version of the robot application.

$sel:robotSoftwareSuite:DescribeRobotApplicationResponse', describeRobotApplicationResponse_robotSoftwareSuite - The robot software suite (ROS distribution) used by the robot application.

$sel:revisionId:DescribeRobotApplicationResponse', describeRobotApplicationResponse_revisionId - The revision id of the robot application.

$sel:tags:DescribeRobotApplicationResponse', describeRobotApplicationResponse_tags - The list of all tags added to the specified robot application.

$sel:httpStatus:DescribeRobotApplicationResponse', describeRobotApplicationResponse_httpStatus - The response's http status code.

Response Lenses

describeRobotApplicationResponse_lastUpdatedAt :: Lens' DescribeRobotApplicationResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the robot application was last updated.

describeRobotApplicationResponse_arn :: Lens' DescribeRobotApplicationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the robot application.

describeRobotApplicationResponse_environment :: Lens' DescribeRobotApplicationResponse (Maybe Environment)

The object that contains the Docker image URI used to create the robot application.

describeRobotApplicationResponse_sources :: Lens' DescribeRobotApplicationResponse (Maybe [Source])

The sources of the robot application.

describeRobotApplicationResponse_name :: Lens' DescribeRobotApplicationResponse (Maybe Text)

The name of the robot application.

describeRobotApplicationResponse_imageDigest :: Lens' DescribeRobotApplicationResponse (Maybe Text)

A SHA256 identifier for the Docker image that you use for your robot application.

describeRobotApplicationResponse_version :: Lens' DescribeRobotApplicationResponse (Maybe Text)

The version of the robot application.

describeRobotApplicationResponse_robotSoftwareSuite :: Lens' DescribeRobotApplicationResponse (Maybe RobotSoftwareSuite)

The robot software suite (ROS distribution) used by the robot application.

describeRobotApplicationResponse_revisionId :: Lens' DescribeRobotApplicationResponse (Maybe Text)

The revision id of the robot application.

describeRobotApplicationResponse_tags :: Lens' DescribeRobotApplicationResponse (Maybe (HashMap Text Text))

The list of all tags added to the specified robot application.

describeRobotApplicationResponse_httpStatus :: Lens' DescribeRobotApplicationResponse Int

The response's http status code.