Network.AWS.RobOMaker.DescribeSimulationApplication

Description

Describes a simulation application.

Creating a Request

data DescribeSimulationApplication

See: newDescribeSimulationApplication smart constructor.

Constructors

DescribeSimulationApplication'

Fields

applicationVersion :: Maybe Text

The version of the simulation application to describe.

application :: Text

The application information for the simulation application.

newDescribeSimulationApplication

Arguments

:: Text
$sel:application:DescribeSimulationApplication'

-> DescribeSimulationApplication

Create a value of DescribeSimulationApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationVersion:DescribeSimulationApplication', describeSimulationApplication_applicationVersion - The version of the simulation application to describe.

$sel:application:DescribeSimulationApplication', describeSimulationApplication_application - The application information for the simulation application.

Request Lenses

describeSimulationApplication_applicationVersion :: Lens' DescribeSimulationApplication (Maybe Text)

The version of the simulation application to describe.

describeSimulationApplication_application :: Lens' DescribeSimulationApplication Text

The application information for the simulation application.

Destructuring the Response

data DescribeSimulationApplicationResponse

See: newDescribeSimulationApplicationResponse smart constructor.

Constructors

DescribeSimulationApplicationResponse'

Fields

renderingEngine :: Maybe RenderingEngine

The rendering engine for the simulation application.

lastUpdatedAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the simulation application was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the robot simulation application.

environment :: Maybe Environment

The object that contains the Docker image URI used to create the simulation application.

sources :: Maybe [Source]

The sources of the simulation application.

name :: Maybe Text

The name of the simulation application.

imageDigest :: Maybe Text

A SHA256 identifier for the Docker image that you use for your simulation application.

version :: Maybe Text

The version of the simulation application.

simulationSoftwareSuite :: Maybe SimulationSoftwareSuite

The simulation software suite used by the simulation application.

robotSoftwareSuite :: Maybe RobotSoftwareSuite

Information about the robot software suite (ROS distribution).

revisionId :: Maybe Text

The revision id of the simulation application.

tags :: Maybe (HashMap Text Text)

The list of all tags added to the specified simulation application.

httpStatus :: Int

The response's http status code.

newDescribeSimulationApplicationResponse

Arguments

:: Int
$sel:httpStatus:DescribeSimulationApplicationResponse'

-> DescribeSimulationApplicationResponse

Create a value of DescribeSimulationApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:renderingEngine:DescribeSimulationApplicationResponse', describeSimulationApplicationResponse_renderingEngine - The rendering engine for the simulation application.

$sel:lastUpdatedAt:DescribeSimulationApplicationResponse', describeSimulationApplicationResponse_lastUpdatedAt - The time, in milliseconds since the epoch, when the simulation application was last updated.

$sel:arn:DescribeSimulationApplicationResponse', describeSimulationApplicationResponse_arn - The Amazon Resource Name (ARN) of the robot simulation application.

$sel:environment:DescribeSimulationApplicationResponse', describeSimulationApplicationResponse_environment - The object that contains the Docker image URI used to create the simulation application.

$sel:sources:DescribeSimulationApplicationResponse', describeSimulationApplicationResponse_sources - The sources of the simulation application.

$sel:name:DescribeSimulationApplicationResponse', describeSimulationApplicationResponse_name - The name of the simulation application.

$sel:imageDigest:DescribeSimulationApplicationResponse', describeSimulationApplicationResponse_imageDigest - A SHA256 identifier for the Docker image that you use for your simulation application.

$sel:version:DescribeSimulationApplicationResponse', describeSimulationApplicationResponse_version - The version of the simulation application.

$sel:simulationSoftwareSuite:DescribeSimulationApplicationResponse', describeSimulationApplicationResponse_simulationSoftwareSuite - The simulation software suite used by the simulation application.

$sel:robotSoftwareSuite:DescribeSimulationApplicationResponse', describeSimulationApplicationResponse_robotSoftwareSuite - Information about the robot software suite (ROS distribution).

$sel:revisionId:DescribeSimulationApplicationResponse', describeSimulationApplicationResponse_revisionId - The revision id of the simulation application.

$sel:tags:DescribeSimulationApplicationResponse', describeSimulationApplicationResponse_tags - The list of all tags added to the specified simulation application.

$sel:httpStatus:DescribeSimulationApplicationResponse', describeSimulationApplicationResponse_httpStatus - The response's http status code.

Response Lenses

describeSimulationApplicationResponse_renderingEngine :: Lens' DescribeSimulationApplicationResponse (Maybe RenderingEngine)

The rendering engine for the simulation application.

describeSimulationApplicationResponse_lastUpdatedAt :: Lens' DescribeSimulationApplicationResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the simulation application was last updated.

describeSimulationApplicationResponse_arn :: Lens' DescribeSimulationApplicationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the robot simulation application.

describeSimulationApplicationResponse_environment :: Lens' DescribeSimulationApplicationResponse (Maybe Environment)

The object that contains the Docker image URI used to create the simulation application.

describeSimulationApplicationResponse_sources :: Lens' DescribeSimulationApplicationResponse (Maybe [Source])

The sources of the simulation application.

describeSimulationApplicationResponse_name :: Lens' DescribeSimulationApplicationResponse (Maybe Text)

The name of the simulation application.

describeSimulationApplicationResponse_imageDigest :: Lens' DescribeSimulationApplicationResponse (Maybe Text)

A SHA256 identifier for the Docker image that you use for your simulation application.

describeSimulationApplicationResponse_version :: Lens' DescribeSimulationApplicationResponse (Maybe Text)

The version of the simulation application.

describeSimulationApplicationResponse_simulationSoftwareSuite :: Lens' DescribeSimulationApplicationResponse (Maybe SimulationSoftwareSuite)

The simulation software suite used by the simulation application.

describeSimulationApplicationResponse_robotSoftwareSuite :: Lens' DescribeSimulationApplicationResponse (Maybe RobotSoftwareSuite)

Information about the robot software suite (ROS distribution).

describeSimulationApplicationResponse_revisionId :: Lens' DescribeSimulationApplicationResponse (Maybe Text)

The revision id of the simulation application.

describeSimulationApplicationResponse_tags :: Lens' DescribeSimulationApplicationResponse (Maybe (HashMap Text Text))

The list of all tags added to the specified simulation application.

describeSimulationApplicationResponse_httpStatus :: Lens' DescribeSimulationApplicationResponse Int

The response's http status code.