Network.AWS.RobOMaker.DescribeSimulationJob

Description

Describes a simulation job.

Creating a Request

data DescribeSimulationJob

See: newDescribeSimulationJob smart constructor.

Constructors

DescribeSimulationJob'

Fields

job :: Text

The Amazon Resource Name (ARN) of the simulation job to be described.

newDescribeSimulationJob

Arguments

:: Text
$sel:job:DescribeSimulationJob'

-> DescribeSimulationJob

Create a value of DescribeSimulationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:job:DescribeSimulationJob', describeSimulationJob_job - The Amazon Resource Name (ARN) of the simulation job to be described.

Request Lenses

describeSimulationJob_job :: Lens' DescribeSimulationJob Text

The Amazon Resource Name (ARN) of the simulation job to be described.

Destructuring the Response

data DescribeSimulationJobResponse

See: newDescribeSimulationJobResponse smart constructor.

Constructors

DescribeSimulationJobResponse'

Fields

failureReason :: Maybe Text

Details about why the simulation job failed. For more information about troubleshooting, see Troubleshooting.

failureBehavior :: Maybe FailureBehavior

The failure behavior for the simulation job.

status :: Maybe SimulationJobStatus

The status of the simulation job.

lastUpdatedAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the simulation job was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the simulation job.

robotApplications :: Maybe (NonEmpty RobotApplicationConfig)

A list of robot applications.

failureCode :: Maybe SimulationJobErrorCode

The failure code of the simulation job if it failed:

InternalServiceError

Internal service error.

RobotApplicationCrash

Robot application exited abnormally.

SimulationApplicationCrash

Simulation application exited abnormally.

BadPermissionsRobotApplication

Robot application bundle could not be downloaded.

BadPermissionsSimulationApplication

Simulation application bundle could not be downloaded.

BadPermissionsS3Output

Unable to publish outputs to customer-provided S3 bucket.

BadPermissionsCloudwatchLogs

Unable to publish logs to customer-provided CloudWatch Logs resource.

SubnetIpLimitExceeded

Subnet IP limit exceeded.

ENILimitExceeded

ENI limit exceeded.

BadPermissionsUserCredentials

Unable to use the Role provided.

InvalidBundleRobotApplication

Robot bundle cannot be extracted (invalid format, bundling error, or other issue).

InvalidBundleSimulationApplication

Simulation bundle cannot be extracted (invalid format, bundling error, or other issue).

RobotApplicationVersionMismatchedEtag

Etag for RobotApplication does not match value during version creation.

SimulationApplicationVersionMismatchedEtag

Etag for SimulationApplication does not match value during version creation.

compute :: Maybe ComputeResponse

Compute information for the simulation job.

networkInterface :: Maybe NetworkInterface

The network interface information for the simulation job.

dataSources :: Maybe [DataSource]

The data sources for the simulation job.

name :: Maybe Text

The name of the simulation job.

vpcConfig :: Maybe VPCConfigResponse

The VPC configuration.

outputLocation :: Maybe OutputLocation

Location for output files generated by the simulation job.

simulationApplications :: Maybe (NonEmpty SimulationApplicationConfig)

A list of simulation applications.

simulationTimeMillis :: Maybe Integer

The simulation job execution duration in milliseconds.

clientRequestToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

lastStartedAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the simulation job was last started.

loggingConfig :: Maybe LoggingConfig

The logging configuration.

iamRole :: Maybe Text

The IAM role that allows the simulation instance to call the AWS APIs that are specified in its associated policies on your behalf.

maxJobDurationInSeconds :: Maybe Integer

The maximum job duration in seconds. The value must be 8 days (691,200 seconds) or less.

tags :: Maybe (HashMap Text Text)

The list of all tags added to the specified simulation job.

httpStatus :: Int

The response's http status code.

newDescribeSimulationJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeSimulationJobResponse'

-> DescribeSimulationJobResponse

Create a value of DescribeSimulationJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:DescribeSimulationJobResponse', describeSimulationJobResponse_failureReason - Details about why the simulation job failed. For more information about troubleshooting, see Troubleshooting.

$sel:failureBehavior:DescribeSimulationJobResponse', describeSimulationJobResponse_failureBehavior - The failure behavior for the simulation job.

$sel:status:DescribeSimulationJobResponse', describeSimulationJobResponse_status - The status of the simulation job.

$sel:lastUpdatedAt:DescribeSimulationJobResponse', describeSimulationJobResponse_lastUpdatedAt - The time, in milliseconds since the epoch, when the simulation job was last updated.

$sel:arn:DescribeSimulationJobResponse', describeSimulationJobResponse_arn - The Amazon Resource Name (ARN) of the simulation job.

$sel:robotApplications:DescribeSimulationJobResponse', describeSimulationJobResponse_robotApplications - A list of robot applications.

$sel:failureCode:DescribeSimulationJobResponse', describeSimulationJobResponse_failureCode - The failure code of the simulation job if it failed:

InternalServiceError

Internal service error.

RobotApplicationCrash

Robot application exited abnormally.

SimulationApplicationCrash

Simulation application exited abnormally.

BadPermissionsRobotApplication

Robot application bundle could not be downloaded.

BadPermissionsSimulationApplication

Simulation application bundle could not be downloaded.

BadPermissionsS3Output

Unable to publish outputs to customer-provided S3 bucket.

BadPermissionsCloudwatchLogs

Unable to publish logs to customer-provided CloudWatch Logs resource.

SubnetIpLimitExceeded

Subnet IP limit exceeded.

ENILimitExceeded

ENI limit exceeded.

BadPermissionsUserCredentials

Unable to use the Role provided.

InvalidBundleRobotApplication

Robot bundle cannot be extracted (invalid format, bundling error, or other issue).

InvalidBundleSimulationApplication

Simulation bundle cannot be extracted (invalid format, bundling error, or other issue).

RobotApplicationVersionMismatchedEtag

Etag for RobotApplication does not match value during version creation.

SimulationApplicationVersionMismatchedEtag

Etag for SimulationApplication does not match value during version creation.

$sel:compute:DescribeSimulationJobResponse', describeSimulationJobResponse_compute - Compute information for the simulation job.

$sel:networkInterface:DescribeSimulationJobResponse', describeSimulationJobResponse_networkInterface - The network interface information for the simulation job.

$sel:dataSources:DescribeSimulationJobResponse', describeSimulationJobResponse_dataSources - The data sources for the simulation job.

$sel:name:DescribeSimulationJobResponse', describeSimulationJobResponse_name - The name of the simulation job.

$sel:vpcConfig:DescribeSimulationJobResponse', describeSimulationJobResponse_vpcConfig - The VPC configuration.

$sel:outputLocation:DescribeSimulationJobResponse', describeSimulationJobResponse_outputLocation - Location for output files generated by the simulation job.

$sel:simulationApplications:DescribeSimulationJobResponse', describeSimulationJobResponse_simulationApplications - A list of simulation applications.

$sel:simulationTimeMillis:DescribeSimulationJobResponse', describeSimulationJobResponse_simulationTimeMillis - The simulation job execution duration in milliseconds.

$sel:clientRequestToken:DescribeSimulationJobResponse', describeSimulationJobResponse_clientRequestToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:lastStartedAt:DescribeSimulationJobResponse', describeSimulationJobResponse_lastStartedAt - The time, in milliseconds since the epoch, when the simulation job was last started.

$sel:loggingConfig:DescribeSimulationJobResponse', describeSimulationJobResponse_loggingConfig - The logging configuration.

$sel:iamRole:DescribeSimulationJobResponse', describeSimulationJobResponse_iamRole - The IAM role that allows the simulation instance to call the AWS APIs that are specified in its associated policies on your behalf.

$sel:maxJobDurationInSeconds:DescribeSimulationJobResponse', describeSimulationJobResponse_maxJobDurationInSeconds - The maximum job duration in seconds. The value must be 8 days (691,200 seconds) or less.

$sel:tags:DescribeSimulationJobResponse', describeSimulationJobResponse_tags - The list of all tags added to the specified simulation job.

$sel:httpStatus:DescribeSimulationJobResponse', describeSimulationJobResponse_httpStatus - The response's http status code.

Response Lenses

describeSimulationJobResponse_failureReason :: Lens' DescribeSimulationJobResponse (Maybe Text)

Details about why the simulation job failed. For more information about troubleshooting, see Troubleshooting.

describeSimulationJobResponse_failureBehavior :: Lens' DescribeSimulationJobResponse (Maybe FailureBehavior)

The failure behavior for the simulation job.

describeSimulationJobResponse_status :: Lens' DescribeSimulationJobResponse (Maybe SimulationJobStatus)

The status of the simulation job.

describeSimulationJobResponse_lastUpdatedAt :: Lens' DescribeSimulationJobResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the simulation job was last updated.

describeSimulationJobResponse_arn :: Lens' DescribeSimulationJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the simulation job.

describeSimulationJobResponse_robotApplications :: Lens' DescribeSimulationJobResponse (Maybe (NonEmpty RobotApplicationConfig))

A list of robot applications.

describeSimulationJobResponse_failureCode :: Lens' DescribeSimulationJobResponse (Maybe SimulationJobErrorCode)

The failure code of the simulation job if it failed:

InternalServiceError

Internal service error.

RobotApplicationCrash

Robot application exited abnormally.

SimulationApplicationCrash

Simulation application exited abnormally.

BadPermissionsRobotApplication

Robot application bundle could not be downloaded.

BadPermissionsSimulationApplication

Simulation application bundle could not be downloaded.

BadPermissionsS3Output

Unable to publish outputs to customer-provided S3 bucket.

BadPermissionsCloudwatchLogs

Unable to publish logs to customer-provided CloudWatch Logs resource.

SubnetIpLimitExceeded

Subnet IP limit exceeded.

ENILimitExceeded

ENI limit exceeded.

BadPermissionsUserCredentials

Unable to use the Role provided.

InvalidBundleRobotApplication

Robot bundle cannot be extracted (invalid format, bundling error, or other issue).

InvalidBundleSimulationApplication

Simulation bundle cannot be extracted (invalid format, bundling error, or other issue).

RobotApplicationVersionMismatchedEtag

Etag for RobotApplication does not match value during version creation.

SimulationApplicationVersionMismatchedEtag

Etag for SimulationApplication does not match value during version creation.

describeSimulationJobResponse_compute :: Lens' DescribeSimulationJobResponse (Maybe ComputeResponse)

Compute information for the simulation job.

describeSimulationJobResponse_networkInterface :: Lens' DescribeSimulationJobResponse (Maybe NetworkInterface)

The network interface information for the simulation job.

describeSimulationJobResponse_dataSources :: Lens' DescribeSimulationJobResponse (Maybe [DataSource])

The data sources for the simulation job.

describeSimulationJobResponse_name :: Lens' DescribeSimulationJobResponse (Maybe Text)

The name of the simulation job.

describeSimulationJobResponse_vpcConfig :: Lens' DescribeSimulationJobResponse (Maybe VPCConfigResponse)

The VPC configuration.

describeSimulationJobResponse_outputLocation :: Lens' DescribeSimulationJobResponse (Maybe OutputLocation)

Location for output files generated by the simulation job.

describeSimulationJobResponse_simulationApplications :: Lens' DescribeSimulationJobResponse (Maybe (NonEmpty SimulationApplicationConfig))

A list of simulation applications.

describeSimulationJobResponse_simulationTimeMillis :: Lens' DescribeSimulationJobResponse (Maybe Integer)

The simulation job execution duration in milliseconds.

describeSimulationJobResponse_clientRequestToken :: Lens' DescribeSimulationJobResponse (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

describeSimulationJobResponse_lastStartedAt :: Lens' DescribeSimulationJobResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the simulation job was last started.

describeSimulationJobResponse_loggingConfig :: Lens' DescribeSimulationJobResponse (Maybe LoggingConfig)

The logging configuration.

describeSimulationJobResponse_iamRole :: Lens' DescribeSimulationJobResponse (Maybe Text)

The IAM role that allows the simulation instance to call the AWS APIs that are specified in its associated policies on your behalf.

describeSimulationJobResponse_maxJobDurationInSeconds :: Lens' DescribeSimulationJobResponse (Maybe Integer)

The maximum job duration in seconds. The value must be 8 days (691,200 seconds) or less.

describeSimulationJobResponse_tags :: Lens' DescribeSimulationJobResponse (Maybe (HashMap Text Text))

The list of all tags added to the specified simulation job.

describeSimulationJobResponse_httpStatus :: Lens' DescribeSimulationJobResponse Int

The response's http status code.