Network.AWS.RobOMaker.ListDeploymentJobs

Description

Returns a list of deployment jobs for a fleet. You can optionally provide filters to retrieve specific deployment jobs.

This operation returns paginated results.

Creating a Request

data ListDeploymentJobs

See: newListDeploymentJobs smart constructor.

Constructors

ListDeploymentJobs'

Fields

filters :: Maybe (NonEmpty Filter)

Optional filters to limit results.

The filter names status and fleetName are supported. When filtering, you must use the complete value of the filtered item. You can use up to three filters, but they must be for the same named item. For example, if you are looking for items with the status InProgress or the status Pending.

nextToken :: Maybe Text

If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListDeploymentJobs again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

maxResults :: Maybe Int

When this parameter is used, ListDeploymentJobs only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListDeploymentJobs request with the returned nextToken value. This value can be between 1 and 200. If this parameter is not used, then ListDeploymentJobs returns up to 200 results and a nextToken value if applicable.

newListDeploymentJobs :: ListDeploymentJobs

Create a value of ListDeploymentJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListDeploymentJobs', listDeploymentJobs_filters - Optional filters to limit results.

The filter names status and fleetName are supported. When filtering, you must use the complete value of the filtered item. You can use up to three filters, but they must be for the same named item. For example, if you are looking for items with the status InProgress or the status Pending.

$sel:nextToken:ListDeploymentJobs', listDeploymentJobs_nextToken - If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListDeploymentJobs again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

$sel:maxResults:ListDeploymentJobs', listDeploymentJobs_maxResults - When this parameter is used, ListDeploymentJobs only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListDeploymentJobs request with the returned nextToken value. This value can be between 1 and 200. If this parameter is not used, then ListDeploymentJobs returns up to 200 results and a nextToken value if applicable.

Request Lenses

listDeploymentJobs_filters :: Lens' ListDeploymentJobs (Maybe (NonEmpty Filter))

Optional filters to limit results.

The filter names status and fleetName are supported. When filtering, you must use the complete value of the filtered item. You can use up to three filters, but they must be for the same named item. For example, if you are looking for items with the status InProgress or the status Pending.

listDeploymentJobs_nextToken :: Lens' ListDeploymentJobs (Maybe Text)

If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListDeploymentJobs again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

listDeploymentJobs_maxResults :: Lens' ListDeploymentJobs (Maybe Int)

When this parameter is used, ListDeploymentJobs only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListDeploymentJobs request with the returned nextToken value. This value can be between 1 and 200. If this parameter is not used, then ListDeploymentJobs returns up to 200 results and a nextToken value if applicable.

Destructuring the Response

data ListDeploymentJobsResponse

See: newListDeploymentJobsResponse smart constructor.

Constructors

ListDeploymentJobsResponse'

Fields

deploymentJobs :: Maybe [DeploymentJob]

A list of deployment jobs that meet the criteria of the request.

nextToken :: Maybe Text

If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListDeploymentJobs again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

httpStatus :: Int

The response's http status code.

newListDeploymentJobsResponse

Arguments

:: Int
$sel:httpStatus:ListDeploymentJobsResponse'

-> ListDeploymentJobsResponse

Create a value of ListDeploymentJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentJobs:ListDeploymentJobsResponse', listDeploymentJobsResponse_deploymentJobs - A list of deployment jobs that meet the criteria of the request.

$sel:nextToken:ListDeploymentJobs', listDeploymentJobsResponse_nextToken - If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListDeploymentJobs again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

$sel:httpStatus:ListDeploymentJobsResponse', listDeploymentJobsResponse_httpStatus - The response's http status code.

Response Lenses

listDeploymentJobsResponse_deploymentJobs :: Lens' ListDeploymentJobsResponse (Maybe [DeploymentJob])

A list of deployment jobs that meet the criteria of the request.

listDeploymentJobsResponse_nextToken :: Lens' ListDeploymentJobsResponse (Maybe Text)

If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListDeploymentJobs again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

listDeploymentJobsResponse_httpStatus :: Lens' ListDeploymentJobsResponse Int

The response's http status code.