Network.AWS.RobOMaker.ListRobots

Description

Returns a list of robots. You can optionally provide filters to retrieve specific robots.

This operation returns paginated results.

Creating a Request

data ListRobots

See: newListRobots smart constructor.

Constructors

ListRobots'

Fields

filters :: Maybe (NonEmpty Filter)

Optional filters to limit results.

The filter names status and fleetName are supported. When filtering, you must use the complete value of the filtered item. You can use up to three filters, but they must be for the same named item. For example, if you are looking for items with the status Registered or the status Available.

nextToken :: Maybe Text

If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListRobots again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

maxResults :: Maybe Int

When this parameter is used, ListRobots only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListRobots request with the returned nextToken value. This value can be between 1 and 200. If this parameter is not used, then ListRobots returns up to 200 results and a nextToken value if applicable.

newListRobots :: ListRobots

Create a value of ListRobots with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListRobots', listRobots_filters - Optional filters to limit results.

The filter names status and fleetName are supported. When filtering, you must use the complete value of the filtered item. You can use up to three filters, but they must be for the same named item. For example, if you are looking for items with the status Registered or the status Available.

$sel:nextToken:ListRobots', listRobots_nextToken - If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListRobots again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

$sel:maxResults:ListRobots', listRobots_maxResults - When this parameter is used, ListRobots only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListRobots request with the returned nextToken value. This value can be between 1 and 200. If this parameter is not used, then ListRobots returns up to 200 results and a nextToken value if applicable.

Request Lenses

listRobots_filters :: Lens' ListRobots (Maybe (NonEmpty Filter))

Optional filters to limit results.

The filter names status and fleetName are supported. When filtering, you must use the complete value of the filtered item. You can use up to three filters, but they must be for the same named item. For example, if you are looking for items with the status Registered or the status Available.

listRobots_nextToken :: Lens' ListRobots (Maybe Text)

If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListRobots again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

listRobots_maxResults :: Lens' ListRobots (Maybe Int)

When this parameter is used, ListRobots only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListRobots request with the returned nextToken value. This value can be between 1 and 200. If this parameter is not used, then ListRobots returns up to 200 results and a nextToken value if applicable.

Destructuring the Response

data ListRobotsResponse

See: newListRobotsResponse smart constructor.

Constructors

ListRobotsResponse'

Fields

robots :: Maybe [Robot]

A list of robots that meet the criteria of the request.

nextToken :: Maybe Text

If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListRobots again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

httpStatus :: Int

The response's http status code.

newListRobotsResponse

Arguments

:: Int
$sel:httpStatus:ListRobotsResponse'

-> ListRobotsResponse

Create a value of ListRobotsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:robots:ListRobotsResponse', listRobotsResponse_robots - A list of robots that meet the criteria of the request.

$sel:nextToken:ListRobots', listRobotsResponse_nextToken - If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListRobots again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

$sel:httpStatus:ListRobotsResponse', listRobotsResponse_httpStatus - The response's http status code.

Response Lenses

listRobotsResponse_robots :: Lens' ListRobotsResponse (Maybe [Robot])

A list of robots that meet the criteria of the request.

listRobotsResponse_nextToken :: Lens' ListRobotsResponse (Maybe Text)

If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListRobots again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

listRobotsResponse_httpStatus :: Lens' ListRobotsResponse Int

The response's http status code.