Network.AWS.RobOMaker.ListWorldGenerationJobs

Description

Lists world generator jobs.

This operation returns paginated results.

Creating a Request

data ListWorldGenerationJobs

See: newListWorldGenerationJobs smart constructor.

Constructors

ListWorldGenerationJobs'

Fields

filters :: Maybe (NonEmpty Filter)

Optional filters to limit results. You can use status and templateId.

nextToken :: Maybe Text

If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListWorldGenerationJobsRequest again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

maxResults :: Maybe Int

When this parameter is used, ListWorldGeneratorJobs only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListWorldGeneratorJobs request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListWorldGeneratorJobs returns up to 100 results and a nextToken value if applicable.

newListWorldGenerationJobs :: ListWorldGenerationJobs

Create a value of ListWorldGenerationJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListWorldGenerationJobs', listWorldGenerationJobs_filters - Optional filters to limit results. You can use status and templateId.

$sel:nextToken:ListWorldGenerationJobs', listWorldGenerationJobs_nextToken - If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListWorldGenerationJobsRequest again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

$sel:maxResults:ListWorldGenerationJobs', listWorldGenerationJobs_maxResults - When this parameter is used, ListWorldGeneratorJobs only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListWorldGeneratorJobs request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListWorldGeneratorJobs returns up to 100 results and a nextToken value if applicable.

Request Lenses

listWorldGenerationJobs_filters :: Lens' ListWorldGenerationJobs (Maybe (NonEmpty Filter))

Optional filters to limit results. You can use status and templateId.

listWorldGenerationJobs_nextToken :: Lens' ListWorldGenerationJobs (Maybe Text)

If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListWorldGenerationJobsRequest again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

listWorldGenerationJobs_maxResults :: Lens' ListWorldGenerationJobs (Maybe Int)

When this parameter is used, ListWorldGeneratorJobs only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListWorldGeneratorJobs request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListWorldGeneratorJobs returns up to 100 results and a nextToken value if applicable.

Destructuring the Response

data ListWorldGenerationJobsResponse

See: newListWorldGenerationJobsResponse smart constructor.

Constructors

ListWorldGenerationJobsResponse'

Fields

nextToken :: Maybe Text

If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListWorldGeneratorJobsRequest again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

httpStatus :: Int

The response's http status code.

worldGenerationJobSummaries :: [WorldGenerationJobSummary]

Summary information for world generator jobs.

newListWorldGenerationJobsResponse

Arguments

:: Int
$sel:httpStatus:ListWorldGenerationJobsResponse'

-> ListWorldGenerationJobsResponse

Create a value of ListWorldGenerationJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListWorldGenerationJobs', listWorldGenerationJobsResponse_nextToken - If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListWorldGeneratorJobsRequest again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

$sel:httpStatus:ListWorldGenerationJobsResponse', listWorldGenerationJobsResponse_httpStatus - The response's http status code.

$sel:worldGenerationJobSummaries:ListWorldGenerationJobsResponse', listWorldGenerationJobsResponse_worldGenerationJobSummaries - Summary information for world generator jobs.

Response Lenses

listWorldGenerationJobsResponse_nextToken :: Lens' ListWorldGenerationJobsResponse (Maybe Text)

If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListWorldGeneratorJobsRequest again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

listWorldGenerationJobsResponse_httpStatus :: Lens' ListWorldGenerationJobsResponse Int

The response's http status code.

listWorldGenerationJobsResponse_worldGenerationJobSummaries :: Lens' ListWorldGenerationJobsResponse [WorldGenerationJobSummary]

Summary information for world generator jobs.