Network.AWS.RobOMaker.SyncDeploymentJob

Description

Syncrhonizes robots in a fleet to the latest deployment. This is helpful if robots were added after a deployment.

Creating a Request

data SyncDeploymentJob

See: newSyncDeploymentJob smart constructor.

Constructors

SyncDeploymentJob'

Fields

clientRequestToken :: Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

fleet :: Text

The target fleet for the synchronization.

newSyncDeploymentJob

Arguments

:: Text
$sel:clientRequestToken:SyncDeploymentJob'

-> Text
$sel:fleet:SyncDeploymentJob'

-> SyncDeploymentJob

Create a value of SyncDeploymentJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientRequestToken:SyncDeploymentJob', syncDeploymentJob_clientRequestToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:fleet:SyncDeploymentJob', syncDeploymentJob_fleet - The target fleet for the synchronization.

Request Lenses

syncDeploymentJob_clientRequestToken :: Lens' SyncDeploymentJob Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

syncDeploymentJob_fleet :: Lens' SyncDeploymentJob Text

The target fleet for the synchronization.

Destructuring the Response

data SyncDeploymentJobResponse

See: newSyncDeploymentJobResponse smart constructor.

Constructors

SyncDeploymentJobResponse'

Fields

failureReason :: Maybe Text

The failure reason if the job fails.

status :: Maybe DeploymentStatus

The status of the synchronization job.

deploymentApplicationConfigs :: Maybe (NonEmpty DeploymentApplicationConfig)

Information about the deployment application configurations.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the synchronization request.

createdAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the fleet was created.

failureCode :: Maybe DeploymentJobErrorCode

The failure code if the job fails:

InternalServiceError

Internal service error.

RobotApplicationCrash

Robot application exited abnormally.

SimulationApplicationCrash

Simulation application exited abnormally.

BadPermissionsRobotApplication

Robot application bundle could not be downloaded.

BadPermissionsSimulationApplication

Simulation application bundle could not be downloaded.

BadPermissionsS3Output

Unable to publish outputs to customer-provided S3 bucket.

BadPermissionsCloudwatchLogs

Unable to publish logs to customer-provided CloudWatch Logs resource.

SubnetIpLimitExceeded

Subnet IP limit exceeded.

ENILimitExceeded

ENI limit exceeded.

BadPermissionsUserCredentials

Unable to use the Role provided.

InvalidBundleRobotApplication

Robot bundle cannot be extracted (invalid format, bundling error, or other issue).

InvalidBundleSimulationApplication

Simulation bundle cannot be extracted (invalid format, bundling error, or other issue).

RobotApplicationVersionMismatchedEtag

Etag for RobotApplication does not match value during version creation.

SimulationApplicationVersionMismatchedEtag

Etag for SimulationApplication does not match value during version creation.

deploymentConfig :: Maybe DeploymentConfig

Information about the deployment configuration.

fleet :: Maybe Text

The Amazon Resource Name (ARN) of the fleet.

httpStatus :: Int

The response's http status code.

newSyncDeploymentJobResponse

Arguments

:: Int
$sel:httpStatus:SyncDeploymentJobResponse'

-> SyncDeploymentJobResponse

Create a value of SyncDeploymentJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:SyncDeploymentJobResponse', syncDeploymentJobResponse_failureReason - The failure reason if the job fails.

$sel:status:SyncDeploymentJobResponse', syncDeploymentJobResponse_status - The status of the synchronization job.

$sel:deploymentApplicationConfigs:SyncDeploymentJobResponse', syncDeploymentJobResponse_deploymentApplicationConfigs - Information about the deployment application configurations.

$sel:arn:SyncDeploymentJobResponse', syncDeploymentJobResponse_arn - The Amazon Resource Name (ARN) of the synchronization request.

$sel:createdAt:SyncDeploymentJobResponse', syncDeploymentJobResponse_createdAt - The time, in milliseconds since the epoch, when the fleet was created.

$sel:failureCode:SyncDeploymentJobResponse', syncDeploymentJobResponse_failureCode - The failure code if the job fails:

InternalServiceError

Internal service error.

RobotApplicationCrash

Robot application exited abnormally.

SimulationApplicationCrash

Simulation application exited abnormally.

BadPermissionsRobotApplication

Robot application bundle could not be downloaded.

BadPermissionsSimulationApplication

Simulation application bundle could not be downloaded.

BadPermissionsS3Output

Unable to publish outputs to customer-provided S3 bucket.

BadPermissionsCloudwatchLogs

Unable to publish logs to customer-provided CloudWatch Logs resource.

SubnetIpLimitExceeded

Subnet IP limit exceeded.

ENILimitExceeded

ENI limit exceeded.

BadPermissionsUserCredentials

Unable to use the Role provided.

InvalidBundleRobotApplication

Robot bundle cannot be extracted (invalid format, bundling error, or other issue).

InvalidBundleSimulationApplication

Simulation bundle cannot be extracted (invalid format, bundling error, or other issue).

RobotApplicationVersionMismatchedEtag

Etag for RobotApplication does not match value during version creation.

SimulationApplicationVersionMismatchedEtag

Etag for SimulationApplication does not match value during version creation.

$sel:deploymentConfig:SyncDeploymentJobResponse', syncDeploymentJobResponse_deploymentConfig - Information about the deployment configuration.

$sel:fleet:SyncDeploymentJob', syncDeploymentJobResponse_fleet - The Amazon Resource Name (ARN) of the fleet.

$sel:httpStatus:SyncDeploymentJobResponse', syncDeploymentJobResponse_httpStatus - The response's http status code.

Response Lenses

syncDeploymentJobResponse_failureReason :: Lens' SyncDeploymentJobResponse (Maybe Text)

The failure reason if the job fails.

syncDeploymentJobResponse_status :: Lens' SyncDeploymentJobResponse (Maybe DeploymentStatus)

The status of the synchronization job.

syncDeploymentJobResponse_deploymentApplicationConfigs :: Lens' SyncDeploymentJobResponse (Maybe (NonEmpty DeploymentApplicationConfig))

Information about the deployment application configurations.

syncDeploymentJobResponse_arn :: Lens' SyncDeploymentJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the synchronization request.

syncDeploymentJobResponse_createdAt :: Lens' SyncDeploymentJobResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the fleet was created.

syncDeploymentJobResponse_failureCode :: Lens' SyncDeploymentJobResponse (Maybe DeploymentJobErrorCode)

The failure code if the job fails:

InternalServiceError

Internal service error.

RobotApplicationCrash

Robot application exited abnormally.

SimulationApplicationCrash

Simulation application exited abnormally.

BadPermissionsRobotApplication

Robot application bundle could not be downloaded.

BadPermissionsSimulationApplication

Simulation application bundle could not be downloaded.

BadPermissionsS3Output

Unable to publish outputs to customer-provided S3 bucket.

BadPermissionsCloudwatchLogs

Unable to publish logs to customer-provided CloudWatch Logs resource.

SubnetIpLimitExceeded

Subnet IP limit exceeded.

ENILimitExceeded

ENI limit exceeded.

BadPermissionsUserCredentials

Unable to use the Role provided.

InvalidBundleRobotApplication

Robot bundle cannot be extracted (invalid format, bundling error, or other issue).

InvalidBundleSimulationApplication

Simulation bundle cannot be extracted (invalid format, bundling error, or other issue).

RobotApplicationVersionMismatchedEtag

Etag for RobotApplication does not match value during version creation.

SimulationApplicationVersionMismatchedEtag

Etag for SimulationApplication does not match value during version creation.

syncDeploymentJobResponse_deploymentConfig :: Lens' SyncDeploymentJobResponse (Maybe DeploymentConfig)

Information about the deployment configuration.

syncDeploymentJobResponse_fleet :: Lens' SyncDeploymentJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the fleet.

syncDeploymentJobResponse_httpStatus :: Lens' SyncDeploymentJobResponse Int

The response's http status code.