Network.AWS.RobOMaker.Types.BatchPolicy

Description

Documentation

data BatchPolicy

Information about the batch policy.

See: newBatchPolicy smart constructor.

Constructors

BatchPolicy'

Fields

timeoutInSeconds :: Maybe Integer

The amount of time, in seconds, to wait for the batch to complete.

If a batch times out, and there are pending requests that were failing due to an internal failure (like InternalServiceError), they will be moved to the failed list and the batch status will be Failed. If the pending requests were failing for any other reason, the failed pending requests will be moved to the failed list and the batch status will be TimedOut.

maxConcurrency :: Maybe Int

The number of active simulation jobs create as part of the batch that can be in an active state at the same time.

Active states include: Pending,Preparing, Running, Restarting, RunningFailed and Terminating. All other states are terminal states.

newBatchPolicy :: BatchPolicy

Create a value of BatchPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeoutInSeconds:BatchPolicy', batchPolicy_timeoutInSeconds - The amount of time, in seconds, to wait for the batch to complete.

If a batch times out, and there are pending requests that were failing due to an internal failure (like InternalServiceError), they will be moved to the failed list and the batch status will be Failed. If the pending requests were failing for any other reason, the failed pending requests will be moved to the failed list and the batch status will be TimedOut.

$sel:maxConcurrency:BatchPolicy', batchPolicy_maxConcurrency - The number of active simulation jobs create as part of the batch that can be in an active state at the same time.

Active states include: Pending,Preparing, Running, Restarting, RunningFailed and Terminating. All other states are terminal states.

batchPolicy_timeoutInSeconds :: Lens' BatchPolicy (Maybe Integer)

The amount of time, in seconds, to wait for the batch to complete.

If a batch times out, and there are pending requests that were failing due to an internal failure (like InternalServiceError), they will be moved to the failed list and the batch status will be Failed. If the pending requests were failing for any other reason, the failed pending requests will be moved to the failed list and the batch status will be TimedOut.

batchPolicy_maxConcurrency :: Lens' BatchPolicy (Maybe Int)

The number of active simulation jobs create as part of the batch that can be in an active state at the same time.

Active states include: Pending,Preparing, Running, Restarting, RunningFailed and Terminating. All other states are terminal states.