Network.AWS.RobOMaker.Types.DeploymentConfig

Description

Documentation

data DeploymentConfig

Information about a deployment configuration.

See: newDeploymentConfig smart constructor.

Constructors

DeploymentConfig'

Fields

concurrentDeploymentPercentage :: Maybe Natural

The percentage of robots receiving the deployment at the same time.

downloadConditionFile :: Maybe S3Object

The download condition file.

failureThresholdPercentage :: Maybe Natural

The percentage of deployments that need to fail before stopping deployment.

robotDeploymentTimeoutInSeconds :: Maybe Integer

The amount of time, in seconds, to wait for deployment to a single robot to complete. Choose a time between 1 minute and 7 days. The default is 5 hours.

newDeploymentConfig :: DeploymentConfig

Create a value of DeploymentConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:concurrentDeploymentPercentage:DeploymentConfig', deploymentConfig_concurrentDeploymentPercentage - The percentage of robots receiving the deployment at the same time.

$sel:downloadConditionFile:DeploymentConfig', deploymentConfig_downloadConditionFile - The download condition file.

$sel:failureThresholdPercentage:DeploymentConfig', deploymentConfig_failureThresholdPercentage - The percentage of deployments that need to fail before stopping deployment.

$sel:robotDeploymentTimeoutInSeconds:DeploymentConfig', deploymentConfig_robotDeploymentTimeoutInSeconds - The amount of time, in seconds, to wait for deployment to a single robot to complete. Choose a time between 1 minute and 7 days. The default is 5 hours.

deploymentConfig_concurrentDeploymentPercentage :: Lens' DeploymentConfig (Maybe Natural)

The percentage of robots receiving the deployment at the same time.

deploymentConfig_downloadConditionFile :: Lens' DeploymentConfig (Maybe S3Object)

The download condition file.

deploymentConfig_failureThresholdPercentage :: Lens' DeploymentConfig (Maybe Natural)

The percentage of deployments that need to fail before stopping deployment.

deploymentConfig_robotDeploymentTimeoutInSeconds :: Lens' DeploymentConfig (Maybe Integer)

The amount of time, in seconds, to wait for deployment to a single robot to complete. Choose a time between 1 minute and 7 days. The default is 5 hours.