Network.AWS.RobOMaker.Types.DeploymentJob

Description

Documentation

data DeploymentJob

Information about a deployment job.

See: newDeploymentJob smart constructor.

Constructors

DeploymentJob'

Fields

failureReason :: Maybe Text

A short description of the reason why the deployment job failed.

status :: Maybe DeploymentStatus

The status of the deployment job.

deploymentApplicationConfigs :: Maybe (NonEmpty DeploymentApplicationConfig)

The deployment application configuration.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the deployment job.

createdAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the deployment job was created.

failureCode :: Maybe DeploymentJobErrorCode

The deployment job failure code.

deploymentConfig :: Maybe DeploymentConfig

The deployment configuration.

fleet :: Maybe Text

The Amazon Resource Name (ARN) of the fleet.

newDeploymentJob :: DeploymentJob

Create a value of DeploymentJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:DeploymentJob', deploymentJob_failureReason - A short description of the reason why the deployment job failed.

$sel:status:DeploymentJob', deploymentJob_status - The status of the deployment job.

$sel:deploymentApplicationConfigs:DeploymentJob', deploymentJob_deploymentApplicationConfigs - The deployment application configuration.

$sel:arn:DeploymentJob', deploymentJob_arn - The Amazon Resource Name (ARN) of the deployment job.

$sel:createdAt:DeploymentJob', deploymentJob_createdAt - The time, in milliseconds since the epoch, when the deployment job was created.

$sel:failureCode:DeploymentJob', deploymentJob_failureCode - The deployment job failure code.

$sel:deploymentConfig:DeploymentJob', deploymentJob_deploymentConfig - The deployment configuration.

$sel:fleet:DeploymentJob', deploymentJob_fleet - The Amazon Resource Name (ARN) of the fleet.

deploymentJob_failureReason :: Lens' DeploymentJob (Maybe Text)

A short description of the reason why the deployment job failed.

deploymentJob_status :: Lens' DeploymentJob (Maybe DeploymentStatus)

The status of the deployment job.

deploymentJob_deploymentApplicationConfigs :: Lens' DeploymentJob (Maybe (NonEmpty DeploymentApplicationConfig))

The deployment application configuration.

deploymentJob_arn :: Lens' DeploymentJob (Maybe Text)

The Amazon Resource Name (ARN) of the deployment job.

deploymentJob_createdAt :: Lens' DeploymentJob (Maybe UTCTime)

The time, in milliseconds since the epoch, when the deployment job was created.

deploymentJob_failureCode :: Lens' DeploymentJob (Maybe DeploymentJobErrorCode)

The deployment job failure code.

deploymentJob_deploymentConfig :: Lens' DeploymentJob (Maybe DeploymentConfig)

The deployment configuration.

deploymentJob_fleet :: Lens' DeploymentJob (Maybe Text)

The Amazon Resource Name (ARN) of the fleet.