Network.AWS.RobOMaker.Types.FailedCreateSimulationJobRequest

Description

Documentation

data FailedCreateSimulationJobRequest

Information about a failed create simulation job request.

See: newFailedCreateSimulationJobRequest smart constructor.

Constructors

FailedCreateSimulationJobRequest'

Fields

failureReason :: Maybe Text

The failure reason of the simulation job request.

failureCode :: Maybe SimulationJobErrorCode

The failure code.

failedAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the simulation job batch failed.

request :: Maybe SimulationJobRequest

The simulation job request.

newFailedCreateSimulationJobRequest :: FailedCreateSimulationJobRequest

Create a value of FailedCreateSimulationJobRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:FailedCreateSimulationJobRequest', failedCreateSimulationJobRequest_failureReason - The failure reason of the simulation job request.

$sel:failureCode:FailedCreateSimulationJobRequest', failedCreateSimulationJobRequest_failureCode - The failure code.

$sel:failedAt:FailedCreateSimulationJobRequest', failedCreateSimulationJobRequest_failedAt - The time, in milliseconds since the epoch, when the simulation job batch failed.

$sel:request:FailedCreateSimulationJobRequest', failedCreateSimulationJobRequest_request - The simulation job request.

failedCreateSimulationJobRequest_failureReason :: Lens' FailedCreateSimulationJobRequest (Maybe Text)

The failure reason of the simulation job request.

failedCreateSimulationJobRequest_failureCode :: Lens' FailedCreateSimulationJobRequest (Maybe SimulationJobErrorCode)

The failure code.

failedCreateSimulationJobRequest_failedAt :: Lens' FailedCreateSimulationJobRequest (Maybe UTCTime)

The time, in milliseconds since the epoch, when the simulation job batch failed.

failedCreateSimulationJobRequest_request :: Lens' FailedCreateSimulationJobRequest (Maybe SimulationJobRequest)

The simulation job request.