Network.AWS.RobOMaker.Types.Robot

Description

Documentation

data Robot

Information about a robot.

See: newRobot smart constructor.

Constructors

Robot'

Fields

lastDeploymentJob :: Maybe Text

The Amazon Resource Name (ARN) of the last deployment job.

status :: Maybe RobotStatus

The status of the robot.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the robot.

createdAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the robot was created.

greenGrassGroupId :: Maybe Text

The Greengrass group associated with the robot.

fleetArn :: Maybe Text

The Amazon Resource Name (ARN) of the fleet.

name :: Maybe Text

The name of the robot.

architecture :: Maybe Architecture

The architecture of the robot.

lastDeploymentTime :: Maybe POSIX

The time of the last deployment.

newRobot :: Robot

Create a value of Robot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastDeploymentJob:Robot', robot_lastDeploymentJob - The Amazon Resource Name (ARN) of the last deployment job.

$sel:status:Robot', robot_status - The status of the robot.

$sel:arn:Robot', robot_arn - The Amazon Resource Name (ARN) of the robot.

$sel:createdAt:Robot', robot_createdAt - The time, in milliseconds since the epoch, when the robot was created.

$sel:greenGrassGroupId:Robot', robot_greenGrassGroupId - The Greengrass group associated with the robot.

$sel:fleetArn:Robot', robot_fleetArn - The Amazon Resource Name (ARN) of the fleet.

$sel:name:Robot', robot_name - The name of the robot.

$sel:architecture:Robot', robot_architecture - The architecture of the robot.

$sel:lastDeploymentTime:Robot', robot_lastDeploymentTime - The time of the last deployment.

robot_lastDeploymentJob :: Lens' Robot (Maybe Text)

The Amazon Resource Name (ARN) of the last deployment job.

robot_status :: Lens' Robot (Maybe RobotStatus)

The status of the robot.

robot_arn :: Lens' Robot (Maybe Text)

The Amazon Resource Name (ARN) of the robot.

robot_createdAt :: Lens' Robot (Maybe UTCTime)

The time, in milliseconds since the epoch, when the robot was created.

robot_greenGrassGroupId :: Lens' Robot (Maybe Text)

The Greengrass group associated with the robot.

robot_fleetArn :: Lens' Robot (Maybe Text)

The Amazon Resource Name (ARN) of the fleet.

robot_name :: Lens' Robot (Maybe Text)

The name of the robot.

robot_architecture :: Lens' Robot (Maybe Architecture)

The architecture of the robot.

robot_lastDeploymentTime :: Lens' Robot (Maybe UTCTime)

The time of the last deployment.