Network.AWS.RobOMaker.Types.RobotDeployment

Description

Documentation

data RobotDeployment

Information about a robot deployment.

See: newRobotDeployment smart constructor.

Constructors

RobotDeployment'

Fields

deploymentStartTime :: Maybe POSIX

The time, in milliseconds since the epoch, when the deployment was started.

failureReason :: Maybe Text

A short description of the reason why the robot deployment failed.

status :: Maybe RobotStatus

The status of the robot deployment.

arn :: Maybe Text

The robot deployment Amazon Resource Name (ARN).

failureCode :: Maybe DeploymentJobErrorCode

The robot deployment failure code.

progressDetail :: Maybe ProgressDetail

Information about how the deployment is progressing.

deploymentFinishTime :: Maybe POSIX

The time, in milliseconds since the epoch, when the deployment finished.

newRobotDeployment :: RobotDeployment

Create a value of RobotDeployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentStartTime:RobotDeployment', robotDeployment_deploymentStartTime - The time, in milliseconds since the epoch, when the deployment was started.

$sel:failureReason:RobotDeployment', robotDeployment_failureReason - A short description of the reason why the robot deployment failed.

$sel:status:RobotDeployment', robotDeployment_status - The status of the robot deployment.

$sel:arn:RobotDeployment', robotDeployment_arn - The robot deployment Amazon Resource Name (ARN).

$sel:failureCode:RobotDeployment', robotDeployment_failureCode - The robot deployment failure code.

$sel:progressDetail:RobotDeployment', robotDeployment_progressDetail - Information about how the deployment is progressing.

$sel:deploymentFinishTime:RobotDeployment', robotDeployment_deploymentFinishTime - The time, in milliseconds since the epoch, when the deployment finished.

robotDeployment_deploymentStartTime :: Lens' RobotDeployment (Maybe UTCTime)

The time, in milliseconds since the epoch, when the deployment was started.

robotDeployment_failureReason :: Lens' RobotDeployment (Maybe Text)

A short description of the reason why the robot deployment failed.

robotDeployment_status :: Lens' RobotDeployment (Maybe RobotStatus)

The status of the robot deployment.

robotDeployment_arn :: Lens' RobotDeployment (Maybe Text)

The robot deployment Amazon Resource Name (ARN).

robotDeployment_failureCode :: Lens' RobotDeployment (Maybe DeploymentJobErrorCode)

The robot deployment failure code.

robotDeployment_progressDetail :: Lens' RobotDeployment (Maybe ProgressDetail)

Information about how the deployment is progressing.

robotDeployment_deploymentFinishTime :: Lens' RobotDeployment (Maybe UTCTime)

The time, in milliseconds since the epoch, when the deployment finished.