Network.AWS.RobOMaker.Types.SimulationApplicationConfig

Description

Documentation

data SimulationApplicationConfig

Information about a simulation application configuration.

See: newSimulationApplicationConfig smart constructor.

Constructors

SimulationApplicationConfig'

Fields

useDefaultUploadConfigurations :: Maybe Bool

A Boolean indicating whether to use default upload configurations. By default, .ros and .gazebo files are uploaded when the application terminates and all ROS topics will be recorded.

If you set this value, you must specify an outputLocation.

useDefaultTools :: Maybe Bool

A Boolean indicating whether to use default simulation application tools. The default tools are rviz, rqt, terminal and rosbag record. The default is False.

applicationVersion :: Maybe Text

The version of the simulation application.

uploadConfigurations :: Maybe [UploadConfiguration]

Information about upload configurations for the simulation application.

tools :: Maybe [Tool]

Information about tools configured for the simulation application.

worldConfigs :: Maybe [WorldConfig]

A list of world configurations.

application :: Text

The application information for the simulation application.

launchConfig :: LaunchConfig

The launch configuration for the simulation application.

newSimulationApplicationConfig

Arguments

:: Text
$sel:application:SimulationApplicationConfig'

-> LaunchConfig
$sel:launchConfig:SimulationApplicationConfig'

-> SimulationApplicationConfig

Create a value of SimulationApplicationConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:useDefaultUploadConfigurations:SimulationApplicationConfig', simulationApplicationConfig_useDefaultUploadConfigurations - A Boolean indicating whether to use default upload configurations. By default, .ros and .gazebo files are uploaded when the application terminates and all ROS topics will be recorded.

If you set this value, you must specify an outputLocation.

$sel:useDefaultTools:SimulationApplicationConfig', simulationApplicationConfig_useDefaultTools - A Boolean indicating whether to use default simulation application tools. The default tools are rviz, rqt, terminal and rosbag record. The default is False.

$sel:applicationVersion:SimulationApplicationConfig', simulationApplicationConfig_applicationVersion - The version of the simulation application.

$sel:uploadConfigurations:SimulationApplicationConfig', simulationApplicationConfig_uploadConfigurations - Information about upload configurations for the simulation application.

$sel:tools:SimulationApplicationConfig', simulationApplicationConfig_tools - Information about tools configured for the simulation application.

$sel:worldConfigs:SimulationApplicationConfig', simulationApplicationConfig_worldConfigs - A list of world configurations.

$sel:application:SimulationApplicationConfig', simulationApplicationConfig_application - The application information for the simulation application.

$sel:launchConfig:SimulationApplicationConfig', simulationApplicationConfig_launchConfig - The launch configuration for the simulation application.

simulationApplicationConfig_useDefaultUploadConfigurations :: Lens' SimulationApplicationConfig (Maybe Bool)

A Boolean indicating whether to use default upload configurations. By default, .ros and .gazebo files are uploaded when the application terminates and all ROS topics will be recorded.

If you set this value, you must specify an outputLocation.

simulationApplicationConfig_useDefaultTools :: Lens' SimulationApplicationConfig (Maybe Bool)

A Boolean indicating whether to use default simulation application tools. The default tools are rviz, rqt, terminal and rosbag record. The default is False.

simulationApplicationConfig_applicationVersion :: Lens' SimulationApplicationConfig (Maybe Text)

The version of the simulation application.

simulationApplicationConfig_uploadConfigurations :: Lens' SimulationApplicationConfig (Maybe [UploadConfiguration])

Information about upload configurations for the simulation application.

simulationApplicationConfig_tools :: Lens' SimulationApplicationConfig (Maybe [Tool])

Information about tools configured for the simulation application.

simulationApplicationConfig_worldConfigs :: Lens' SimulationApplicationConfig (Maybe [WorldConfig])

A list of world configurations.

simulationApplicationConfig_application :: Lens' SimulationApplicationConfig Text

The application information for the simulation application.

simulationApplicationConfig_launchConfig :: Lens' SimulationApplicationConfig LaunchConfig

The launch configuration for the simulation application.