Network.AWS.RobOMaker.Types.SimulationApplicationSummary

Description

Documentation

data SimulationApplicationSummary

Summary information for a simulation application.

See: newSimulationApplicationSummary smart constructor.

Constructors

SimulationApplicationSummary'

Fields

lastUpdatedAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the simulation application was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the simulation application.

name :: Maybe Text

The name of the simulation application.

version :: Maybe Text

The version of the simulation application.

simulationSoftwareSuite :: Maybe SimulationSoftwareSuite

Information about a simulation software suite.

robotSoftwareSuite :: Maybe RobotSoftwareSuite

Information about a robot software suite (ROS distribution).

newSimulationApplicationSummary :: SimulationApplicationSummary

Create a value of SimulationApplicationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedAt:SimulationApplicationSummary', simulationApplicationSummary_lastUpdatedAt - The time, in milliseconds since the epoch, when the simulation application was last updated.

$sel:arn:SimulationApplicationSummary', simulationApplicationSummary_arn - The Amazon Resource Name (ARN) of the simulation application.

$sel:name:SimulationApplicationSummary', simulationApplicationSummary_name - The name of the simulation application.

$sel:version:SimulationApplicationSummary', simulationApplicationSummary_version - The version of the simulation application.

$sel:simulationSoftwareSuite:SimulationApplicationSummary', simulationApplicationSummary_simulationSoftwareSuite - Information about a simulation software suite.

$sel:robotSoftwareSuite:SimulationApplicationSummary', simulationApplicationSummary_robotSoftwareSuite - Information about a robot software suite (ROS distribution).

simulationApplicationSummary_lastUpdatedAt :: Lens' SimulationApplicationSummary (Maybe UTCTime)

The time, in milliseconds since the epoch, when the simulation application was last updated.

simulationApplicationSummary_arn :: Lens' SimulationApplicationSummary (Maybe Text)

The Amazon Resource Name (ARN) of the simulation application.

simulationApplicationSummary_name :: Lens' SimulationApplicationSummary (Maybe Text)

The name of the simulation application.

simulationApplicationSummary_version :: Lens' SimulationApplicationSummary (Maybe Text)

The version of the simulation application.

simulationApplicationSummary_simulationSoftwareSuite :: Lens' SimulationApplicationSummary (Maybe SimulationSoftwareSuite)

Information about a simulation software suite.

simulationApplicationSummary_robotSoftwareSuite :: Lens' SimulationApplicationSummary (Maybe RobotSoftwareSuite)

Information about a robot software suite (ROS distribution).