Network.AWS.RobOMaker.Types.SimulationJob

Description

Documentation

data SimulationJob

Information about a simulation job.

See: newSimulationJob smart constructor.

Constructors

SimulationJob'

Fields

failureReason :: Maybe Text

The reason why the simulation job failed.

failureBehavior :: Maybe FailureBehavior

The failure behavior the simulation job.

Continue

Leaves the host running for its maximum timeout duration after a 4XX error code.

Fail

Stop the simulation job and terminate the instance.

status :: Maybe SimulationJobStatus

Status of the simulation job.

lastUpdatedAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the simulation job was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the simulation job.

robotApplications :: Maybe (NonEmpty RobotApplicationConfig)

A list of robot applications.

failureCode :: Maybe SimulationJobErrorCode

The failure code of the simulation job if it failed.

compute :: Maybe ComputeResponse

Compute information for the simulation job

networkInterface :: Maybe NetworkInterface

Information about a network interface.

dataSources :: Maybe [DataSource]

The data sources for the simulation job.

name :: Maybe Text

The name of the simulation job.

vpcConfig :: Maybe VPCConfigResponse

VPC configuration information.

outputLocation :: Maybe OutputLocation

Location for output files generated by the simulation job.

simulationApplications :: Maybe (NonEmpty SimulationApplicationConfig)

A list of simulation applications.

simulationTimeMillis :: Maybe Integer

The simulation job execution duration in milliseconds.

clientRequestToken :: Maybe Text

A unique identifier for this SimulationJob request.

lastStartedAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the simulation job was last started.

loggingConfig :: Maybe LoggingConfig

The logging configuration.

iamRole :: Maybe Text

The IAM role that allows the simulation instance to call the AWS APIs that are specified in its associated policies on your behalf. This is how credentials are passed in to your simulation job.

maxJobDurationInSeconds :: Maybe Integer

The maximum simulation job duration in seconds. The value must be 8 days (691,200 seconds) or less.

tags :: Maybe (HashMap Text Text)

A map that contains tag keys and tag values that are attached to the simulation job.

newSimulationJob :: SimulationJob

Create a value of SimulationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:SimulationJob', simulationJob_failureReason - The reason why the simulation job failed.

$sel:failureBehavior:SimulationJob', simulationJob_failureBehavior - The failure behavior the simulation job.

Continue

Leaves the host running for its maximum timeout duration after a 4XX error code.

Fail

Stop the simulation job and terminate the instance.

$sel:status:SimulationJob', simulationJob_status - Status of the simulation job.

$sel:lastUpdatedAt:SimulationJob', simulationJob_lastUpdatedAt - The time, in milliseconds since the epoch, when the simulation job was last updated.

$sel:arn:SimulationJob', simulationJob_arn - The Amazon Resource Name (ARN) of the simulation job.

$sel:robotApplications:SimulationJob', simulationJob_robotApplications - A list of robot applications.

$sel:failureCode:SimulationJob', simulationJob_failureCode - The failure code of the simulation job if it failed.

$sel:compute:SimulationJob', simulationJob_compute - Compute information for the simulation job

$sel:networkInterface:SimulationJob', simulationJob_networkInterface - Information about a network interface.

$sel:dataSources:SimulationJob', simulationJob_dataSources - The data sources for the simulation job.

$sel:name:SimulationJob', simulationJob_name - The name of the simulation job.

$sel:vpcConfig:SimulationJob', simulationJob_vpcConfig - VPC configuration information.

$sel:outputLocation:SimulationJob', simulationJob_outputLocation - Location for output files generated by the simulation job.

$sel:simulationApplications:SimulationJob', simulationJob_simulationApplications - A list of simulation applications.

$sel:simulationTimeMillis:SimulationJob', simulationJob_simulationTimeMillis - The simulation job execution duration in milliseconds.

$sel:clientRequestToken:SimulationJob', simulationJob_clientRequestToken - A unique identifier for this SimulationJob request.

$sel:lastStartedAt:SimulationJob', simulationJob_lastStartedAt - The time, in milliseconds since the epoch, when the simulation job was last started.

$sel:loggingConfig:SimulationJob', simulationJob_loggingConfig - The logging configuration.

$sel:iamRole:SimulationJob', simulationJob_iamRole - The IAM role that allows the simulation instance to call the AWS APIs that are specified in its associated policies on your behalf. This is how credentials are passed in to your simulation job.

$sel:maxJobDurationInSeconds:SimulationJob', simulationJob_maxJobDurationInSeconds - The maximum simulation job duration in seconds. The value must be 8 days (691,200 seconds) or less.

$sel:tags:SimulationJob', simulationJob_tags - A map that contains tag keys and tag values that are attached to the simulation job.

simulationJob_failureReason :: Lens' SimulationJob (Maybe Text)

The reason why the simulation job failed.

simulationJob_failureBehavior :: Lens' SimulationJob (Maybe FailureBehavior)

The failure behavior the simulation job.

Continue

Leaves the host running for its maximum timeout duration after a 4XX error code.

Fail

Stop the simulation job and terminate the instance.

simulationJob_status :: Lens' SimulationJob (Maybe SimulationJobStatus)

Status of the simulation job.

simulationJob_lastUpdatedAt :: Lens' SimulationJob (Maybe UTCTime)

The time, in milliseconds since the epoch, when the simulation job was last updated.

simulationJob_arn :: Lens' SimulationJob (Maybe Text)

The Amazon Resource Name (ARN) of the simulation job.

simulationJob_robotApplications :: Lens' SimulationJob (Maybe (NonEmpty RobotApplicationConfig))

A list of robot applications.

simulationJob_failureCode :: Lens' SimulationJob (Maybe SimulationJobErrorCode)

The failure code of the simulation job if it failed.

simulationJob_compute :: Lens' SimulationJob (Maybe ComputeResponse)

Compute information for the simulation job

simulationJob_networkInterface :: Lens' SimulationJob (Maybe NetworkInterface)

Information about a network interface.

simulationJob_dataSources :: Lens' SimulationJob (Maybe [DataSource])

The data sources for the simulation job.

simulationJob_name :: Lens' SimulationJob (Maybe Text)

The name of the simulation job.

simulationJob_vpcConfig :: Lens' SimulationJob (Maybe VPCConfigResponse)

VPC configuration information.

simulationJob_outputLocation :: Lens' SimulationJob (Maybe OutputLocation)

Location for output files generated by the simulation job.

simulationJob_simulationApplications :: Lens' SimulationJob (Maybe (NonEmpty SimulationApplicationConfig))

A list of simulation applications.

simulationJob_simulationTimeMillis :: Lens' SimulationJob (Maybe Integer)

The simulation job execution duration in milliseconds.

simulationJob_clientRequestToken :: Lens' SimulationJob (Maybe Text)

A unique identifier for this SimulationJob request.

simulationJob_lastStartedAt :: Lens' SimulationJob (Maybe UTCTime)

The time, in milliseconds since the epoch, when the simulation job was last started.

simulationJob_loggingConfig :: Lens' SimulationJob (Maybe LoggingConfig)

The logging configuration.

simulationJob_iamRole :: Lens' SimulationJob (Maybe Text)

The IAM role that allows the simulation instance to call the AWS APIs that are specified in its associated policies on your behalf. This is how credentials are passed in to your simulation job.

simulationJob_maxJobDurationInSeconds :: Lens' SimulationJob (Maybe Integer)

The maximum simulation job duration in seconds. The value must be 8 days (691,200 seconds) or less.

simulationJob_tags :: Lens' SimulationJob (Maybe (HashMap Text Text))

A map that contains tag keys and tag values that are attached to the simulation job.