Network.AWS.RobOMaker.Types.SimulationJobBatchSummary

Description

Documentation

data SimulationJobBatchSummary

Information about a simulation job batch.

See: newSimulationJobBatchSummary smart constructor.

Constructors

SimulationJobBatchSummary'

Fields

status :: Maybe SimulationJobBatchStatus

The status of the simulation job batch.

Pending

The simulation job batch request is pending.

InProgress

The simulation job batch is in progress.

Failed

The simulation job batch failed. One or more simulation job requests could not be completed due to an internal failure (like InternalServiceError). See failureCode and failureReason for more information.

Completed

The simulation batch job completed. A batch is complete when (1) there are no pending simulation job requests in the batch and none of the failed simulation job requests are due to InternalServiceError and (2) when all created simulation jobs have reached a terminal state (for example, Completed or Failed).

Canceled

The simulation batch job was cancelled.

Canceling

The simulation batch job is being cancelled.

Completing

The simulation batch job is completing.

TimingOut

The simulation job batch is timing out.
If a batch timing out, and there are pending requests that were failing due to an internal failure (like InternalServiceError), the batch status will be Failed. If there are no such failing request, the batch status will be TimedOut.

TimedOut

The simulation batch job timed out.

createdRequestCount :: Maybe Int

The number of created simulation job requests.

lastUpdatedAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the simulation job batch was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the batch.

createdAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the simulation job batch was created.

pendingRequestCount :: Maybe Int

The number of pending simulation job requests.

failedRequestCount :: Maybe Int

The number of failed simulation job requests.

newSimulationJobBatchSummary :: SimulationJobBatchSummary

Create a value of SimulationJobBatchSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SimulationJobBatchSummary', simulationJobBatchSummary_status - The status of the simulation job batch.

Pending

The simulation job batch request is pending.

InProgress

The simulation job batch is in progress.

Failed

The simulation job batch failed. One or more simulation job requests could not be completed due to an internal failure (like InternalServiceError). See failureCode and failureReason for more information.

Completed

The simulation batch job completed. A batch is complete when (1) there are no pending simulation job requests in the batch and none of the failed simulation job requests are due to InternalServiceError and (2) when all created simulation jobs have reached a terminal state (for example, Completed or Failed).

Canceled

The simulation batch job was cancelled.

Canceling

The simulation batch job is being cancelled.

Completing

The simulation batch job is completing.

TimingOut

The simulation job batch is timing out.
If a batch timing out, and there are pending requests that were failing due to an internal failure (like InternalServiceError), the batch status will be Failed. If there are no such failing request, the batch status will be TimedOut.

TimedOut

The simulation batch job timed out.

$sel:createdRequestCount:SimulationJobBatchSummary', simulationJobBatchSummary_createdRequestCount - The number of created simulation job requests.

$sel:lastUpdatedAt:SimulationJobBatchSummary', simulationJobBatchSummary_lastUpdatedAt - The time, in milliseconds since the epoch, when the simulation job batch was last updated.

$sel:arn:SimulationJobBatchSummary', simulationJobBatchSummary_arn - The Amazon Resource Name (ARN) of the batch.

$sel:createdAt:SimulationJobBatchSummary', simulationJobBatchSummary_createdAt - The time, in milliseconds since the epoch, when the simulation job batch was created.

$sel:pendingRequestCount:SimulationJobBatchSummary', simulationJobBatchSummary_pendingRequestCount - The number of pending simulation job requests.

$sel:failedRequestCount:SimulationJobBatchSummary', simulationJobBatchSummary_failedRequestCount - The number of failed simulation job requests.

simulationJobBatchSummary_status :: Lens' SimulationJobBatchSummary (Maybe SimulationJobBatchStatus)

The status of the simulation job batch.

Pending

The simulation job batch request is pending.

InProgress

The simulation job batch is in progress.

Failed

The simulation job batch failed. One or more simulation job requests could not be completed due to an internal failure (like InternalServiceError). See failureCode and failureReason for more information.

Completed

The simulation batch job completed. A batch is complete when (1) there are no pending simulation job requests in the batch and none of the failed simulation job requests are due to InternalServiceError and (2) when all created simulation jobs have reached a terminal state (for example, Completed or Failed).

Canceled

The simulation batch job was cancelled.

Canceling

The simulation batch job is being cancelled.

Completing

The simulation batch job is completing.

TimingOut

The simulation job batch is timing out.
If a batch timing out, and there are pending requests that were failing due to an internal failure (like InternalServiceError), the batch status will be Failed. If there are no such failing request, the batch status will be TimedOut.

TimedOut

The simulation batch job timed out.

simulationJobBatchSummary_createdRequestCount :: Lens' SimulationJobBatchSummary (Maybe Int)

The number of created simulation job requests.

simulationJobBatchSummary_lastUpdatedAt :: Lens' SimulationJobBatchSummary (Maybe UTCTime)

The time, in milliseconds since the epoch, when the simulation job batch was last updated.

simulationJobBatchSummary_arn :: Lens' SimulationJobBatchSummary (Maybe Text)

The Amazon Resource Name (ARN) of the batch.

simulationJobBatchSummary_createdAt :: Lens' SimulationJobBatchSummary (Maybe UTCTime)

The time, in milliseconds since the epoch, when the simulation job batch was created.

simulationJobBatchSummary_pendingRequestCount :: Lens' SimulationJobBatchSummary (Maybe Int)

The number of pending simulation job requests.

simulationJobBatchSummary_failedRequestCount :: Lens' SimulationJobBatchSummary (Maybe Int)

The number of failed simulation job requests.