Network.AWS.RobOMaker.Types.SimulationJobRequest

Description

Documentation

data SimulationJobRequest

Information about a simulation job request.

See: newSimulationJobRequest smart constructor.

Constructors

SimulationJobRequest'

Fields

failureBehavior :: Maybe FailureBehavior

The failure behavior the simulation job.

Continue

Leaves the host running for its maximum timeout duration after a 4XX error code.

Fail

Stop the simulation job and terminate the instance.

robotApplications :: Maybe (NonEmpty RobotApplicationConfig)

The robot applications to use in the simulation job.

compute :: Maybe Compute

Compute information for the simulation job

dataSources :: Maybe (NonEmpty DataSourceConfig)

Specify data sources to mount read-only files from S3 into your simulation. These files are available under /opt/robomaker/datasources/data_source_name.

There is a limit of 100 files and a combined size of 25GB for all DataSourceConfig objects.

useDefaultApplications :: Maybe Bool

A Boolean indicating whether to use default applications in the simulation job. Default applications include Gazebo, rqt, rviz and terminal access.

vpcConfig :: Maybe VPCConfig

outputLocation :: Maybe OutputLocation

simulationApplications :: Maybe (NonEmpty SimulationApplicationConfig)

The simulation applications to use in the simulation job.

loggingConfig :: Maybe LoggingConfig

iamRole :: Maybe Text

The IAM role name that allows the simulation instance to call the AWS APIs that are specified in its associated policies on your behalf. This is how credentials are passed in to your simulation job.

tags :: Maybe (HashMap Text Text)

A map that contains tag keys and tag values that are attached to the simulation job request.

maxJobDurationInSeconds :: Integer

The maximum simulation job duration in seconds. The value must be 8 days (691,200 seconds) or less.

newSimulationJobRequest

Arguments

:: Integer
$sel:maxJobDurationInSeconds:SimulationJobRequest'

-> SimulationJobRequest

Create a value of SimulationJobRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureBehavior:SimulationJobRequest', simulationJobRequest_failureBehavior - The failure behavior the simulation job.

Continue

Leaves the host running for its maximum timeout duration after a 4XX error code.

Fail

Stop the simulation job and terminate the instance.

$sel:robotApplications:SimulationJobRequest', simulationJobRequest_robotApplications - The robot applications to use in the simulation job.

$sel:compute:SimulationJobRequest', simulationJobRequest_compute - Compute information for the simulation job

$sel:dataSources:SimulationJobRequest', simulationJobRequest_dataSources - Specify data sources to mount read-only files from S3 into your simulation. These files are available under /opt/robomaker/datasources/data_source_name.

There is a limit of 100 files and a combined size of 25GB for all DataSourceConfig objects.

$sel:useDefaultApplications:SimulationJobRequest', simulationJobRequest_useDefaultApplications - A Boolean indicating whether to use default applications in the simulation job. Default applications include Gazebo, rqt, rviz and terminal access.

$sel:vpcConfig:SimulationJobRequest', simulationJobRequest_vpcConfig - Undocumented member.

$sel:outputLocation:SimulationJobRequest', simulationJobRequest_outputLocation - Undocumented member.

$sel:simulationApplications:SimulationJobRequest', simulationJobRequest_simulationApplications - The simulation applications to use in the simulation job.

$sel:loggingConfig:SimulationJobRequest', simulationJobRequest_loggingConfig - Undocumented member.

$sel:iamRole:SimulationJobRequest', simulationJobRequest_iamRole - The IAM role name that allows the simulation instance to call the AWS APIs that are specified in its associated policies on your behalf. This is how credentials are passed in to your simulation job.

$sel:tags:SimulationJobRequest', simulationJobRequest_tags - A map that contains tag keys and tag values that are attached to the simulation job request.

$sel:maxJobDurationInSeconds:SimulationJobRequest', simulationJobRequest_maxJobDurationInSeconds - The maximum simulation job duration in seconds. The value must be 8 days (691,200 seconds) or less.

simulationJobRequest_failureBehavior :: Lens' SimulationJobRequest (Maybe FailureBehavior)

The failure behavior the simulation job.

Continue

Leaves the host running for its maximum timeout duration after a 4XX error code.

Fail

Stop the simulation job and terminate the instance.

simulationJobRequest_robotApplications :: Lens' SimulationJobRequest (Maybe (NonEmpty RobotApplicationConfig))

The robot applications to use in the simulation job.

simulationJobRequest_compute :: Lens' SimulationJobRequest (Maybe Compute)

Compute information for the simulation job

simulationJobRequest_dataSources :: Lens' SimulationJobRequest (Maybe (NonEmpty DataSourceConfig))

Specify data sources to mount read-only files from S3 into your simulation. These files are available under /opt/robomaker/datasources/data_source_name.

There is a limit of 100 files and a combined size of 25GB for all DataSourceConfig objects.

simulationJobRequest_useDefaultApplications :: Lens' SimulationJobRequest (Maybe Bool)

A Boolean indicating whether to use default applications in the simulation job. Default applications include Gazebo, rqt, rviz and terminal access.

simulationJobRequest_vpcConfig :: Lens' SimulationJobRequest (Maybe VPCConfig)

Undocumented member.

simulationJobRequest_outputLocation :: Lens' SimulationJobRequest (Maybe OutputLocation)

Undocumented member.

simulationJobRequest_simulationApplications :: Lens' SimulationJobRequest (Maybe (NonEmpty SimulationApplicationConfig))

The simulation applications to use in the simulation job.

simulationJobRequest_loggingConfig :: Lens' SimulationJobRequest (Maybe LoggingConfig)

Undocumented member.

simulationJobRequest_iamRole :: Lens' SimulationJobRequest (Maybe Text)

The IAM role name that allows the simulation instance to call the AWS APIs that are specified in its associated policies on your behalf. This is how credentials are passed in to your simulation job.

simulationJobRequest_tags :: Lens' SimulationJobRequest (Maybe (HashMap Text Text))

A map that contains tag keys and tag values that are attached to the simulation job request.

simulationJobRequest_maxJobDurationInSeconds :: Lens' SimulationJobRequest Integer

The maximum simulation job duration in seconds. The value must be 8 days (691,200 seconds) or less.