Network.AWS.RobOMaker.Types.Source

Description

Documentation

data Source

Information about a source.

See: newSource smart constructor.

Constructors

Source'

Fields

etag :: Maybe Text

A hash of the object specified by s3Bucket and s3Key.

s3Key :: Maybe Text

The s3 object key.

architecture :: Maybe Architecture

The taget processor architecture for the application.

s3Bucket :: Maybe Text

The s3 bucket name.

newSource :: Source

Create a value of Source with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:etag:Source', source_etag - A hash of the object specified by s3Bucket and s3Key.

$sel:s3Key:Source', source_s3Key - The s3 object key.

$sel:architecture:Source', source_architecture - The taget processor architecture for the application.

$sel:s3Bucket:Source', source_s3Bucket - The s3 bucket name.

source_etag :: Lens' Source (Maybe Text)

A hash of the object specified by s3Bucket and s3Key.

source_s3Key :: Lens' Source (Maybe Text)

The s3 object key.

source_architecture :: Lens' Source (Maybe Architecture)

The taget processor architecture for the application.

source_s3Bucket :: Lens' Source (Maybe Text)

The s3 bucket name.