Network.AWS.RobOMaker.UpdateSimulationApplication

Description

Updates a simulation application.

Creating a Request

data UpdateSimulationApplication

See: newUpdateSimulationApplication smart constructor.

Constructors

UpdateSimulationApplication'

Fields

renderingEngine :: Maybe RenderingEngine

The rendering engine for the simulation application.

currentRevisionId :: Maybe Text

The revision id for the robot application.

environment :: Maybe Environment

The object that contains the Docker image URI for your simulation application.

sources :: Maybe [SourceConfig]

The sources of the simulation application.

application :: Text

The application information for the simulation application.

simulationSoftwareSuite :: SimulationSoftwareSuite

The simulation software suite used by the simulation application.

robotSoftwareSuite :: RobotSoftwareSuite

Information about the robot software suite (ROS distribution).

newUpdateSimulationApplication

Arguments

:: Text
$sel:application:UpdateSimulationApplication'

-> SimulationSoftwareSuite
$sel:simulationSoftwareSuite:UpdateSimulationApplication'

-> RobotSoftwareSuite
$sel:robotSoftwareSuite:UpdateSimulationApplication'

-> UpdateSimulationApplication

Create a value of UpdateSimulationApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:renderingEngine:UpdateSimulationApplication', updateSimulationApplication_renderingEngine - The rendering engine for the simulation application.

$sel:currentRevisionId:UpdateSimulationApplication', updateSimulationApplication_currentRevisionId - The revision id for the robot application.

$sel:environment:UpdateSimulationApplication', updateSimulationApplication_environment - The object that contains the Docker image URI for your simulation application.

$sel:sources:UpdateSimulationApplication', updateSimulationApplication_sources - The sources of the simulation application.

$sel:application:UpdateSimulationApplication', updateSimulationApplication_application - The application information for the simulation application.

$sel:simulationSoftwareSuite:UpdateSimulationApplication', updateSimulationApplication_simulationSoftwareSuite - The simulation software suite used by the simulation application.

$sel:robotSoftwareSuite:UpdateSimulationApplication', updateSimulationApplication_robotSoftwareSuite - Information about the robot software suite (ROS distribution).

Request Lenses

updateSimulationApplication_renderingEngine :: Lens' UpdateSimulationApplication (Maybe RenderingEngine)

The rendering engine for the simulation application.

updateSimulationApplication_currentRevisionId :: Lens' UpdateSimulationApplication (Maybe Text)

The revision id for the robot application.

updateSimulationApplication_environment :: Lens' UpdateSimulationApplication (Maybe Environment)

The object that contains the Docker image URI for your simulation application.

updateSimulationApplication_sources :: Lens' UpdateSimulationApplication (Maybe [SourceConfig])

The sources of the simulation application.

updateSimulationApplication_application :: Lens' UpdateSimulationApplication Text

The application information for the simulation application.

updateSimulationApplication_simulationSoftwareSuite :: Lens' UpdateSimulationApplication SimulationSoftwareSuite

The simulation software suite used by the simulation application.

updateSimulationApplication_robotSoftwareSuite :: Lens' UpdateSimulationApplication RobotSoftwareSuite

Information about the robot software suite (ROS distribution).

Destructuring the Response

data UpdateSimulationApplicationResponse

See: newUpdateSimulationApplicationResponse smart constructor.

Constructors

UpdateSimulationApplicationResponse'

Fields

renderingEngine :: Maybe RenderingEngine

The rendering engine for the simulation application.

lastUpdatedAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the simulation application was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the updated simulation application.

environment :: Maybe Environment

The object that contains the Docker image URI used for your simulation application.

sources :: Maybe [Source]

The sources of the simulation application.

name :: Maybe Text

The name of the simulation application.

version :: Maybe Text

The version of the robot application.

simulationSoftwareSuite :: Maybe SimulationSoftwareSuite

The simulation software suite used by the simulation application.

robotSoftwareSuite :: Maybe RobotSoftwareSuite

Information about the robot software suite (ROS distribution).

revisionId :: Maybe Text

The revision id of the simulation application.

httpStatus :: Int

The response's http status code.

newUpdateSimulationApplicationResponse

Arguments

:: Int
$sel:httpStatus:UpdateSimulationApplicationResponse'

-> UpdateSimulationApplicationResponse

Create a value of UpdateSimulationApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:renderingEngine:UpdateSimulationApplication', updateSimulationApplicationResponse_renderingEngine - The rendering engine for the simulation application.

$sel:lastUpdatedAt:UpdateSimulationApplicationResponse', updateSimulationApplicationResponse_lastUpdatedAt - The time, in milliseconds since the epoch, when the simulation application was last updated.

$sel:arn:UpdateSimulationApplicationResponse', updateSimulationApplicationResponse_arn - The Amazon Resource Name (ARN) of the updated simulation application.

$sel:environment:UpdateSimulationApplication', updateSimulationApplicationResponse_environment - The object that contains the Docker image URI used for your simulation application.

$sel:sources:UpdateSimulationApplication', updateSimulationApplicationResponse_sources - The sources of the simulation application.

$sel:name:UpdateSimulationApplicationResponse', updateSimulationApplicationResponse_name - The name of the simulation application.

$sel:version:UpdateSimulationApplicationResponse', updateSimulationApplicationResponse_version - The version of the robot application.

$sel:simulationSoftwareSuite:UpdateSimulationApplication', updateSimulationApplicationResponse_simulationSoftwareSuite - The simulation software suite used by the simulation application.

$sel:robotSoftwareSuite:UpdateSimulationApplication', updateSimulationApplicationResponse_robotSoftwareSuite - Information about the robot software suite (ROS distribution).

$sel:revisionId:UpdateSimulationApplicationResponse', updateSimulationApplicationResponse_revisionId - The revision id of the simulation application.

$sel:httpStatus:UpdateSimulationApplicationResponse', updateSimulationApplicationResponse_httpStatus - The response's http status code.

Response Lenses

updateSimulationApplicationResponse_renderingEngine :: Lens' UpdateSimulationApplicationResponse (Maybe RenderingEngine)

The rendering engine for the simulation application.

updateSimulationApplicationResponse_lastUpdatedAt :: Lens' UpdateSimulationApplicationResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the simulation application was last updated.

updateSimulationApplicationResponse_arn :: Lens' UpdateSimulationApplicationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the updated simulation application.

updateSimulationApplicationResponse_environment :: Lens' UpdateSimulationApplicationResponse (Maybe Environment)

The object that contains the Docker image URI used for your simulation application.

updateSimulationApplicationResponse_sources :: Lens' UpdateSimulationApplicationResponse (Maybe [Source])

The sources of the simulation application.

updateSimulationApplicationResponse_name :: Lens' UpdateSimulationApplicationResponse (Maybe Text)

The name of the simulation application.

updateSimulationApplicationResponse_version :: Lens' UpdateSimulationApplicationResponse (Maybe Text)

The version of the robot application.

updateSimulationApplicationResponse_simulationSoftwareSuite :: Lens' UpdateSimulationApplicationResponse (Maybe SimulationSoftwareSuite)

The simulation software suite used by the simulation application.

updateSimulationApplicationResponse_robotSoftwareSuite :: Lens' UpdateSimulationApplicationResponse (Maybe RobotSoftwareSuite)

Information about the robot software suite (ROS distribution).

updateSimulationApplicationResponse_revisionId :: Lens' UpdateSimulationApplicationResponse (Maybe Text)

The revision id of the simulation application.

updateSimulationApplicationResponse_httpStatus :: Lens' UpdateSimulationApplicationResponse Int

The response's http status code.