Network.AWS.Route53

Description

Route 53 is a web service that enables you to manage your DNS records.

Actions on Hosted Zones

POST CreateHostedZone

data CreateHostedZone

Creates a new hosted zone.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_CreateHostedZone.html

Constructors

CreateHostedZone

Fields

chzName :: !Text

The name of the domain.

chzCallerReference :: !CallerReference

A unique string that identifies the request and that allows failed CreateHostedZone requests to be retried without the risk of executing the operation twice.

chzHostedZoneConfig :: Maybe Config

A complex type that contains an optional comment about your hosted zone.

data CreateHostedZoneResponse

Constructors

CreateHostedZoneResponse

Fields

chzrHostedZone :: !HostedZone

Information about the hosted zone.

chzrChangeInfo :: !ChangeInfo

Information about the changes being made to the hosted zone.

chzrDelegationSet :: !DelegationSet

The name servers for this hosted zone.

GET GetHostedZone

newtype GetHostedZone

Gets information about a specified hosted zone.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_GetHostedZone.html

Constructors

GetHostedZone

Fields

ghzId :: HostedZoneId

Hosted Zone Id.

data GetHostedZoneResponse

Constructors

GetHostedZoneResponse

Fields

ghzrHostedZone :: !HostedZone

Information about the hosted zone.

ghzrDelegationSet :: !DelegationSet

The name servers for this hosted zone.

GET ListHostedZones

data ListHostedZones

Gets a list of the hosted zones that are associated with the current AWS account.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_ListHostedZones.html

Constructors

ListHostedZones

Fields

lhzMarker :: Maybe Text

To get the next group of maxitems hosted zones, submit another request with the value of the NextMarker element that was returned in the previous response.

lhzMaxItems :: Maybe Integer

Maximum number of hosted zones to include in the response.

data ListHostedZonesResponse

Constructors

ListHostedZonesResponse

Fields

lhzrHostedZones :: [HostedZone]

A list of hosted zone descriptions.

lhzrIsTruncated :: !Bool

Whether the result list has been truncated.

lhzrMarker :: Maybe Text

Value of the marker in the previous request.

lhzrNextMarker :: Maybe Text

If IsTruncated is true, the hosted zone ID of the first hosted zone in the next group of maxitems hosted zones.

lhzrMaxItems :: !Integer

Value of maxitems in the previous request.

DELETE DeleteHostedZone

newtype DeleteHostedZone

Deletes a hosted zone.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_DeleteHostedZone.html

Constructors

DeleteHostedZone

Fields

dhzId :: HostedZoneId

Hosted Zone Id.

data DeleteHostedZoneResponse

Constructors

DeleteHostedZoneResponse

Fields

dhzrChangeInfo :: !ChangeInfo

Information about the changes being made to the hosted zone.

Actions on Record Sets

POST ChangeResourceRecordSets

data ChangeResourceRecordSets

Adds, deletes, and changes resource record sets in a Route 53 hosted zone.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_ChangeResourceRecordSets.html

Constructors

ChangeResourceRecordSets

Fields

crrsZoneId :: !HostedZoneId

Hosted Zone Id.

crrsChangeBatch :: !ChangeBatch

ChangeBatch describing the record set modifications you wish to perform.

data ChangeResourceRecordSetsResponse

Constructors

ChangeResourceRecordSetsResponse

Fields

crrsrChangeInfo :: !ChangeInfo

GET ListResourceRecordSets

data ListResourceRecordSets

Lists details about all of the resource record sets in a hosted zone.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_ListResourceRecordSets.html

Constructors

ListResourceRecordSets

Fields

lrrsZoneId :: !HostedZoneId

The ID of the hosted zone containing the resource records sets to be retrieved.

lrrsName :: Maybe Text

The first name in the lexicographic ordering of domain names to be retrieved in the response to the ListResourceRecordSets request.

lrrsType :: Maybe RecordType

The type of resource record set to begin the record listing from.

lrrsIdentifier :: Maybe Text

Weighted and latency resource record sets only: If results were truncated for a given DNS name and type, the value of SetIdentifier for the next resource record set that has the current DNS name and type.

lrrsMaxItems :: Maybe Integer

The maximum number of resource records sets to include in the response xml for this request. If the response includes more than maxitems resource record sets, the value of the IsTruncated element in the response is true, and the values of the NextRecordName and NextRecordType elements in the response identify the first resource record set in the next group of maxitems resource record sets.

data ListResourceRecordSetsResponse

Constructors

ListResourceRecordSetsResponse

Fields

lrrsrResourceRecordSets :: [ResourceRecordSet]

A list of resource record sets.

lrrsrIsTruncated :: !Bool

Whether the result list has been truncated.

lrrsrMaxItems :: !Integer

Value of maxitems in the previous request.

lrrsrNextRecordName :: Maybe Text

If IsTruncated is true, the DNS domain name of the first resource record set in the next group of maxitems resource record sets.

lrrsrNextRecordType :: Maybe RecordType

If IsTruncated is true, the DNS record type of the first resource record set in the next group of maxitems resource record sets.

lrrsrNextRecordIdentifier :: Maybe Text

If IsTruncated is true and results were truncated for a weighted, latency, or failover resource record set, the value of SetIdentifier for the first resource record set in the next group of maxitems resource record sets.

GET GetChange

newtype GetChange

Returns the current status of a change batch request that you submitted by using ChangeResourceRecordSets.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_GetChange.html

Constructors

GetChange

Fields

gcId :: ChangeId

The ID of the change batch request. The value that you specify here is the value that POST ChangeResourceRecordSets returned in the Id element when you submitted the request.

data GetChangeResponse

Constructors

GetChangeResponse

Fields

gcrChangeInfo :: !ChangeInfo

A description of the changes being made to the hosted zone.

Actions on Health Checks

POST CreateHealthCheck

data CreateHealthCheck

Creates a new health check.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_CreateHealthCheck.html

Constructors

CreateHealthCheck

Fields

chcCallerReference :: !CallerReference

A unique string that identifies the request and that allows failed CreateHealthCheck requests to be retried without the risk of executing the operation twice. You must use a unique CallerReference string every time you create a health check.

chcHealthCheckConfig :: !HealthCheckConfig

Health check configuration.

data CreateHealthCheckResponse

Constructors

CreateHealthCheckResponse

Fields

chcrHealthCheck :: !HealthCheck

Information about the created health check.

GET GetHealthCheck

newtype GetHealthCheck

Gets information about a specified health check.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_GetHealthCheck.html

Constructors

GetHealthCheck

Fields

ghcId :: HealthCheckId

The ID for the health check for which you want detailed information. When you created the health check, CreateHealthCheck returned the ID in the response, in the HealthCheckId element.

data GetHealthCheckResponse

Constructors

GetHealthCheckResponse

Fields

ghcrHealthCheck :: !HealthCheck

Information about a health check.

GET ListHealthChecks

data ListHealthChecks

Gets a list of the health checks that are associated with the current AWS account.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_ListHealthChecks.html

Constructors

ListHealthChecks

Fields

lhcMarker :: Maybe Text

If the response to a ListHealthChecks is more than one page, marker is the health check ID for the first health check on the next page of results.

lhcMaxItems :: Maybe Integer

The maximum number of HealthCheck elements you want ListHealthChecks to return on each page of the response body. If the AWS account includes more HealthCheck elements than the value of maxitems, the response is broken into pages. Each page contains the number of HealthCheck elements specified by maxitems.

data ListHealthChecksResponse

Constructors

ListHealthChecksResponse

Fields

lhcrHealthChecks :: [HealthCheck]

A list of health check descriptions.

lhcrIsTruncated :: !Bool

Whether the result list has been truncated.

lhcrMarker :: Maybe Text

Value of the marker in the previous request.

lhcrNextMarker :: Maybe Text

If IsTruncated is true, the hosted zone ID of the first hosted zone in the next group of maxitems hosted zones.

lhcrMaxItems :: !Integer

Value of maxitems in the previous request.

DELETE DeleteHealthCheck

newtype DeleteHealthCheck

Deletes a health check.

Caution: Route 53 does not prevent you from deleting a health check even if the health check is associated with one or more resource record sets. If you delete a health check and you don't update the associated resource record sets, the future status of the health check cannot be predicted and may change. This will affect the routing of DNS queries for your DNS failover configuration.

http://docs.aws.amazon.com/Route53/latest/APIReference/API_DeleteHealthCheck.html

Constructors

DeleteHealthCheck

Fields

dhcId :: HealthCheckId

Health Check Id.

data DeleteHealthCheckResponse

Constructors

DeleteHealthCheckResponse

Data Types

route53 :: Service

Currently supported version of the Route53 service.

route53NS :: ByteString

XML namespace to annotate Route53 elements with.

route53Elem :: ByteString -> NName ByteString

Helper to define Route53 namespaced XML elements.

class Prefixed a where

Methods

prefixed :: a -> ByteString

data ErrorType

Constructors

ErrorType

Fields

etType :: !Text

etCode :: !Text

etMessage :: !Text

data ErrorResponse

Constructors

ErrorResponse

Fields

erError :: !ErrorType

erRequestId :: !Text

InvalidChangeBatch

Fields

erMessages :: [Text]

newtype CallerReference

Constructors

CallerReference

Fields

unCallerReference :: Text

callerRef :: IO CallerReference

data Protocol

Constructors

HTTP

TCP

data RecordType

Constructors

A

AAAA

CNAME

MX

NS

PTR

SOA

SPF

SRV

TXT

data Config

Constructors

Config

Fields

cComment :: Maybe Text

newtype HostedZoneId

Constructors

HostedZoneId

Fields

unHostedZoneId :: Text

data HostedZone

Constructors

HostedZone

Fields

hzId :: !HostedZoneId

The ID of the hosted zone.

hzName :: !Text

The name of the domain. For resource record types that include a domain name, specify a fully qualified domain name, for example, www.example.com. The trailing dot is optional; Route

hzCallerReference :: !CallerReference

A unique string that identifies the request to create the hosted zone.

hzConfig :: !Config

A complex type that contains an optional comment about your hosted zone.

hzResourceRecordSetCount :: !Integer

The number of resource record sets in the hosted zone.

newtype DelegationSet

Constructors

DelegationSet

Fields

dsNameServers :: [Text]

newtype ChangeId

Constructors

ChangeId

Fields

unChangeId :: Text

data ChangeStatus

Constructors

PENDING

INSYNC

data ChangeInfo

Constructors

ChangeInfo

Fields

ciId :: !ChangeId

ciStatus :: !ChangeStatus

ciSubmittedAt :: !UTCTime

data ChangeAction

Constructors

CreateAction

DeleteAction

data Change

Constructors

Change

Fields

cAction :: !ChangeAction

cResourceRecordSet :: !ResourceRecordSet

data ChangeBatch

Constructors

ChangeBatch

Fields

cbComment :: Maybe Text

cbChanges :: [Change]

newtype ResourceRecords

Constructors

ResourceRecords

Fields

rrValues :: [Text]

data AliasTarget

Constructors

AliasTarget

Fields

atHostedZoneId :: !HostedZoneId

atDNSName :: !Text

atEvaluateTargetHealth :: !Bool

data Failover

Constructors

PRIMARY

SECONDARY

data ResourceRecordSet

Constructors

FailoverRecordSet

Fields

rrsName :: !Text

rrsType :: !RecordType

rrsSetIdentifier :: !Text

rrsFailover :: !Failover

rrsTTL :: !Integer

rrsResourceRecords :: !ResourceRecords

rrsHealthCheckId :: Maybe HealthCheckId

FailoverAliasRecordSet

Fields

rrsName :: !Text

rrsType :: !RecordType

rrsSetIdentifier :: !Text

rrsFailover :: !Failover

rrsAliasTarget :: !AliasTarget

rrsHealthCheckId :: Maybe HealthCheckId

LatencyRecordSet

Fields

rrsName :: !Text

rrsType :: !RecordType

rrsSetIdentifier :: !Text

rrsRegion :: !Region

rrsTTL :: !Integer

rrsResourceRecords :: !ResourceRecords

rrsHealthCheckId :: Maybe HealthCheckId

LatencyAliasRecordSet

Fields

rrsName :: !Text

rrsType :: !RecordType

rrsSetIdentifier :: !Text

rrsRegion :: !Region

rrsAliasTarget :: !AliasTarget

rrsHealthCheckId :: Maybe HealthCheckId

WeightedRecordSet

Fields

rrsName :: !Text

rrsType :: !RecordType

rrsSetIdentifier :: !Text

rrsWeight :: !Integer

rrsTTL :: !Integer

rrsResourceRecords :: !ResourceRecords

rrsHealthCheckId :: Maybe HealthCheckId

WeightedAliasRecordSet

Fields

rrsName :: !Text

rrsType :: !RecordType

rrsSetIdentifier :: !Text

rrsWeight :: !Integer

rrsAliasTarget :: !AliasTarget

rrsHealthCheckId :: Maybe HealthCheckId

BasicRecordSet

Fields

rrsName :: !Text

rrsType :: !RecordType

rrsTTL :: !Integer

rrsResourceRecords :: !ResourceRecords

rrsHealthCheckId :: Maybe HealthCheckId

AliasRecordSet

Fields

rrsName :: !Text

rrsType :: !RecordType

rrsAliasTarget :: !AliasTarget

rrsHealthCheckId :: Maybe HealthCheckId

data HealthCheckConfig

Constructors

HealthCheckConfig

Fields

hccIPAddress :: !Text

hccPort :: !Int

hccType :: !Protocol

hccResourcePath :: !Text

hccFullyQualifiedDomainName :: !Text

newtype HealthCheckId

Constructors

HealthCheckId

Fields

unHealthCheckId :: Text

data HealthCheck

Constructors

HealthCheck

Fields

hcId :: !HealthCheckId

hcCallerReference :: !CallerReference

hcHealthCheckConfig :: !HealthCheckConfig

Common

module Network.AWS