Network.AWS.Route53.ChangeTagsForResource

Description

Adds, edits, or deletes tags for a health check or a hosted zone.

For information about using tags for cost allocation, see Using Cost Allocation Tags in the Billing and Cost Management User Guide.

Creating a Request

data ChangeTagsForResource

A complex type that contains information about the tags that you want to add, edit, or delete.

See: newChangeTagsForResource smart constructor.

Constructors

ChangeTagsForResource'

Fields

addTags :: Maybe (NonEmpty Tag)

A complex type that contains a list of the tags that you want to add to the specified health check or hosted zone and/or the tags that you want to edit Value for.

You can add a maximum of 10 tags to a health check or a hosted zone.

removeTagKeys :: Maybe (NonEmpty Text)

A complex type that contains a list of the tags that you want to delete from the specified health check or hosted zone. You can specify up to 10 keys.

resourceType :: TagResourceType

The type of the resource.

The resource type for health checks is healthcheck.

The resource type for hosted zones is hostedzone.

resourceId :: Text

The ID of the resource for which you want to add, change, or delete tags.

newChangeTagsForResource

Arguments

:: TagResourceType
$sel:resourceType:ChangeTagsForResource'

-> Text
$sel:resourceId:ChangeTagsForResource'

-> ChangeTagsForResource

Create a value of ChangeTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:addTags:ChangeTagsForResource', changeTagsForResource_addTags - A complex type that contains a list of the tags that you want to add to the specified health check or hosted zone and/or the tags that you want to edit Value for.

You can add a maximum of 10 tags to a health check or a hosted zone.

$sel:removeTagKeys:ChangeTagsForResource', changeTagsForResource_removeTagKeys - A complex type that contains a list of the tags that you want to delete from the specified health check or hosted zone. You can specify up to 10 keys.

$sel:resourceType:ChangeTagsForResource', changeTagsForResource_resourceType - The type of the resource.

The resource type for health checks is healthcheck.

The resource type for hosted zones is hostedzone.

$sel:resourceId:ChangeTagsForResource', changeTagsForResource_resourceId - The ID of the resource for which you want to add, change, or delete tags.

Request Lenses

changeTagsForResource_addTags :: Lens' ChangeTagsForResource (Maybe (NonEmpty Tag))

A complex type that contains a list of the tags that you want to add to the specified health check or hosted zone and/or the tags that you want to edit Value for.

You can add a maximum of 10 tags to a health check or a hosted zone.

changeTagsForResource_removeTagKeys :: Lens' ChangeTagsForResource (Maybe (NonEmpty Text))

A complex type that contains a list of the tags that you want to delete from the specified health check or hosted zone. You can specify up to 10 keys.

changeTagsForResource_resourceType :: Lens' ChangeTagsForResource TagResourceType

The type of the resource.

The resource type for health checks is healthcheck.

The resource type for hosted zones is hostedzone.

changeTagsForResource_resourceId :: Lens' ChangeTagsForResource Text

The ID of the resource for which you want to add, change, or delete tags.

Destructuring the Response

data ChangeTagsForResourceResponse

Empty response for the request.

See: newChangeTagsForResourceResponse smart constructor.

Constructors

ChangeTagsForResourceResponse'

Fields

httpStatus :: Int

The response's http status code.

newChangeTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ChangeTagsForResourceResponse'

-> ChangeTagsForResourceResponse

Create a value of ChangeTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ChangeTagsForResourceResponse', changeTagsForResourceResponse_httpStatus - The response's http status code.

Response Lenses

changeTagsForResourceResponse_httpStatus :: Lens' ChangeTagsForResourceResponse Int

The response's http status code.