Network.AWS.Route53.CreateTrafficPolicyInstance

Description

Creates resource record sets in a specified hosted zone based on the settings in a specified traffic policy version. In addition, CreateTrafficPolicyInstance associates the resource record sets with a specified domain name (such as example.com) or subdomain name (such as www.example.com). Amazon Route 53 responds to DNS queries for the domain or subdomain name by using the resource record sets that CreateTrafficPolicyInstance created.

Creating a Request

data CreateTrafficPolicyInstance

A complex type that contains information about the resource record sets that you want to create based on a specified traffic policy.

See: newCreateTrafficPolicyInstance smart constructor.

Constructors

CreateTrafficPolicyInstance'

Fields

hostedZoneId :: ResourceId

The ID of the hosted zone that you want Amazon Route 53 to create resource record sets in by using the configuration in a traffic policy.

name :: Text

The domain name (such as example.com) or subdomain name (such as www.example.com) for which Amazon Route 53 responds to DNS queries by using the resource record sets that Route 53 creates for this traffic policy instance.

ttl :: Natural

(Optional) The TTL that you want Amazon Route 53 to assign to all of the resource record sets that it creates in the specified hosted zone.

trafficPolicyId :: Text

The ID of the traffic policy that you want to use to create resource record sets in the specified hosted zone.

trafficPolicyVersion :: Natural

The version of the traffic policy that you want to use to create resource record sets in the specified hosted zone.

newCreateTrafficPolicyInstance

Arguments

:: ResourceId
$sel:hostedZoneId:CreateTrafficPolicyInstance'

-> Text
$sel:name:CreateTrafficPolicyInstance'

-> Natural
$sel:ttl:CreateTrafficPolicyInstance'

-> Text
$sel:trafficPolicyId:CreateTrafficPolicyInstance'

-> Natural
$sel:trafficPolicyVersion:CreateTrafficPolicyInstance'

-> CreateTrafficPolicyInstance

Create a value of CreateTrafficPolicyInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostedZoneId:CreateTrafficPolicyInstance', createTrafficPolicyInstance_hostedZoneId - The ID of the hosted zone that you want Amazon Route 53 to create resource record sets in by using the configuration in a traffic policy.

$sel:name:CreateTrafficPolicyInstance', createTrafficPolicyInstance_name - The domain name (such as example.com) or subdomain name (such as www.example.com) for which Amazon Route 53 responds to DNS queries by using the resource record sets that Route 53 creates for this traffic policy instance.

$sel:ttl:CreateTrafficPolicyInstance', createTrafficPolicyInstance_ttl - (Optional) The TTL that you want Amazon Route 53 to assign to all of the resource record sets that it creates in the specified hosted zone.

$sel:trafficPolicyId:CreateTrafficPolicyInstance', createTrafficPolicyInstance_trafficPolicyId - The ID of the traffic policy that you want to use to create resource record sets in the specified hosted zone.

$sel:trafficPolicyVersion:CreateTrafficPolicyInstance', createTrafficPolicyInstance_trafficPolicyVersion - The version of the traffic policy that you want to use to create resource record sets in the specified hosted zone.

Request Lenses

createTrafficPolicyInstance_hostedZoneId :: Lens' CreateTrafficPolicyInstance ResourceId

The ID of the hosted zone that you want Amazon Route 53 to create resource record sets in by using the configuration in a traffic policy.

createTrafficPolicyInstance_name :: Lens' CreateTrafficPolicyInstance Text

The domain name (such as example.com) or subdomain name (such as www.example.com) for which Amazon Route 53 responds to DNS queries by using the resource record sets that Route 53 creates for this traffic policy instance.

createTrafficPolicyInstance_ttl :: Lens' CreateTrafficPolicyInstance Natural

(Optional) The TTL that you want Amazon Route 53 to assign to all of the resource record sets that it creates in the specified hosted zone.

createTrafficPolicyInstance_trafficPolicyId :: Lens' CreateTrafficPolicyInstance Text

The ID of the traffic policy that you want to use to create resource record sets in the specified hosted zone.

createTrafficPolicyInstance_trafficPolicyVersion :: Lens' CreateTrafficPolicyInstance Natural

The version of the traffic policy that you want to use to create resource record sets in the specified hosted zone.

Destructuring the Response

data CreateTrafficPolicyInstanceResponse

A complex type that contains the response information for the CreateTrafficPolicyInstance request.

See: newCreateTrafficPolicyInstanceResponse smart constructor.

Constructors

CreateTrafficPolicyInstanceResponse'

Fields

httpStatus :: Int

The response's http status code.

trafficPolicyInstance :: TrafficPolicyInstance

A complex type that contains settings for the new traffic policy instance.

location :: Text

A unique URL that represents a new traffic policy instance.

newCreateTrafficPolicyInstanceResponse

Arguments

:: Int
$sel:httpStatus:CreateTrafficPolicyInstanceResponse'

-> TrafficPolicyInstance
$sel:trafficPolicyInstance:CreateTrafficPolicyInstanceResponse'

-> Text
$sel:location:CreateTrafficPolicyInstanceResponse'

-> CreateTrafficPolicyInstanceResponse

Create a value of CreateTrafficPolicyInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateTrafficPolicyInstanceResponse', createTrafficPolicyInstanceResponse_httpStatus - The response's http status code.

$sel:trafficPolicyInstance:CreateTrafficPolicyInstanceResponse', createTrafficPolicyInstanceResponse_trafficPolicyInstance - A complex type that contains settings for the new traffic policy instance.

$sel:location:CreateTrafficPolicyInstanceResponse', createTrafficPolicyInstanceResponse_location - A unique URL that represents a new traffic policy instance.

Response Lenses

createTrafficPolicyInstanceResponse_httpStatus :: Lens' CreateTrafficPolicyInstanceResponse Int

The response's http status code.

createTrafficPolicyInstanceResponse_trafficPolicyInstance :: Lens' CreateTrafficPolicyInstanceResponse TrafficPolicyInstance

A complex type that contains settings for the new traffic policy instance.

createTrafficPolicyInstanceResponse_location :: Lens' CreateTrafficPolicyInstanceResponse Text

A unique URL that represents a new traffic policy instance.