Network.AWS.Route53.CreateTrafficPolicyVersion

Description

Creates a new version of an existing traffic policy. When you create a new version of a traffic policy, you specify the ID of the traffic policy that you want to update and a JSON-formatted document that describes the new version. You use traffic policies to create multiple DNS resource record sets for one domain name (such as example.com) or one subdomain name (such as www.example.com). You can create a maximum of 1000 versions of a traffic policy. If you reach the limit and need to create another version, you'll need to start a new traffic policy.

Creating a Request

data CreateTrafficPolicyVersion

A complex type that contains information about the traffic policy that you want to create a new version for.

See: newCreateTrafficPolicyVersion smart constructor.

Constructors

CreateTrafficPolicyVersion'

Fields

comment :: Maybe Text

The comment that you specified in the CreateTrafficPolicyVersion request, if any.

id :: Text

The ID of the traffic policy for which you want to create a new version.

document :: Text

The definition of this version of the traffic policy, in JSON format. You specified the JSON in the CreateTrafficPolicyVersion request. For more information about the JSON format, see CreateTrafficPolicy.

newCreateTrafficPolicyVersion

Arguments

:: Text
$sel:id:CreateTrafficPolicyVersion'

-> Text
$sel:document:CreateTrafficPolicyVersion'

-> CreateTrafficPolicyVersion

Create a value of CreateTrafficPolicyVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:comment:CreateTrafficPolicyVersion', createTrafficPolicyVersion_comment - The comment that you specified in the CreateTrafficPolicyVersion request, if any.

$sel:id:CreateTrafficPolicyVersion', createTrafficPolicyVersion_id - The ID of the traffic policy for which you want to create a new version.

$sel:document:CreateTrafficPolicyVersion', createTrafficPolicyVersion_document - The definition of this version of the traffic policy, in JSON format. You specified the JSON in the CreateTrafficPolicyVersion request. For more information about the JSON format, see CreateTrafficPolicy.

Request Lenses

createTrafficPolicyVersion_comment :: Lens' CreateTrafficPolicyVersion (Maybe Text)

The comment that you specified in the CreateTrafficPolicyVersion request, if any.

createTrafficPolicyVersion_id :: Lens' CreateTrafficPolicyVersion Text

The ID of the traffic policy for which you want to create a new version.

createTrafficPolicyVersion_document :: Lens' CreateTrafficPolicyVersion Text

The definition of this version of the traffic policy, in JSON format. You specified the JSON in the CreateTrafficPolicyVersion request. For more information about the JSON format, see CreateTrafficPolicy.

Destructuring the Response

data CreateTrafficPolicyVersionResponse

A complex type that contains the response information for the CreateTrafficPolicyVersion request.

See: newCreateTrafficPolicyVersionResponse smart constructor.

Constructors

CreateTrafficPolicyVersionResponse'

Fields

httpStatus :: Int

The response's http status code.

trafficPolicy :: TrafficPolicy

A complex type that contains settings for the new version of the traffic policy.

location :: Text

A unique URL that represents a new traffic policy version.

newCreateTrafficPolicyVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateTrafficPolicyVersionResponse'

-> TrafficPolicy
$sel:trafficPolicy:CreateTrafficPolicyVersionResponse'

-> Text
$sel:location:CreateTrafficPolicyVersionResponse'

-> CreateTrafficPolicyVersionResponse

Create a value of CreateTrafficPolicyVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateTrafficPolicyVersionResponse', createTrafficPolicyVersionResponse_httpStatus - The response's http status code.

$sel:trafficPolicy:CreateTrafficPolicyVersionResponse', createTrafficPolicyVersionResponse_trafficPolicy - A complex type that contains settings for the new version of the traffic policy.

$sel:location:CreateTrafficPolicyVersionResponse', createTrafficPolicyVersionResponse_location - A unique URL that represents a new traffic policy version.

Response Lenses

createTrafficPolicyVersionResponse_httpStatus :: Lens' CreateTrafficPolicyVersionResponse Int

The response's http status code.

createTrafficPolicyVersionResponse_trafficPolicy :: Lens' CreateTrafficPolicyVersionResponse TrafficPolicy

A complex type that contains settings for the new version of the traffic policy.

createTrafficPolicyVersionResponse_location :: Lens' CreateTrafficPolicyVersionResponse Text

A unique URL that represents a new traffic policy version.