Network.AWS.Route53.CreateVPCAssociationAuthorization

Description

Authorizes the Amazon Web Services account that created a specified VPC to submit an AssociateVPCWithHostedZone request to associate the VPC with a specified hosted zone that was created by a different account. To submit a CreateVPCAssociationAuthorization request, you must use the account that created the hosted zone. After you authorize the association, use the account that created the VPC to submit an AssociateVPCWithHostedZone request.

If you want to associate multiple VPCs that you created by using one account with a hosted zone that you created by using a different account, you must submit one authorization request for each VPC.

Creating a Request

data CreateVPCAssociationAuthorization

A complex type that contains information about the request to authorize associating a VPC with your private hosted zone. Authorization is only required when a private hosted zone and a VPC were created by using different accounts.

See: newCreateVPCAssociationAuthorization smart constructor.

Constructors

CreateVPCAssociationAuthorization'

Fields

hostedZoneId :: ResourceId

The ID of the private hosted zone that you want to authorize associating a VPC with.

vpc :: VPC

A complex type that contains the VPC ID and region for the VPC that you want to authorize associating with your hosted zone.

newCreateVPCAssociationAuthorization

Arguments

:: ResourceId
$sel:hostedZoneId:CreateVPCAssociationAuthorization'

-> VPC
$sel:vpc:CreateVPCAssociationAuthorization'

-> CreateVPCAssociationAuthorization

Create a value of CreateVPCAssociationAuthorization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostedZoneId:CreateVPCAssociationAuthorization', createVPCAssociationAuthorization_hostedZoneId - The ID of the private hosted zone that you want to authorize associating a VPC with.

$sel:vpc:CreateVPCAssociationAuthorization', createVPCAssociationAuthorization_vpc - A complex type that contains the VPC ID and region for the VPC that you want to authorize associating with your hosted zone.

Request Lenses

createVPCAssociationAuthorization_hostedZoneId :: Lens' CreateVPCAssociationAuthorization ResourceId

The ID of the private hosted zone that you want to authorize associating a VPC with.

createVPCAssociationAuthorization_vpc :: Lens' CreateVPCAssociationAuthorization VPC

A complex type that contains the VPC ID and region for the VPC that you want to authorize associating with your hosted zone.

Destructuring the Response

data CreateVPCAssociationAuthorizationResponse

A complex type that contains the response information from a CreateVPCAssociationAuthorization request.

See: newCreateVPCAssociationAuthorizationResponse smart constructor.

Constructors

CreateVPCAssociationAuthorizationResponse'

Fields

httpStatus :: Int

The response's http status code.

hostedZoneId :: ResourceId

The ID of the hosted zone that you authorized associating a VPC with.

vpc :: VPC

The VPC that you authorized associating with a hosted zone.

newCreateVPCAssociationAuthorizationResponse

Arguments

:: Int
$sel:httpStatus:CreateVPCAssociationAuthorizationResponse'

-> ResourceId
$sel:hostedZoneId:CreateVPCAssociationAuthorization'

-> VPC
$sel:vpc:CreateVPCAssociationAuthorization'

-> CreateVPCAssociationAuthorizationResponse

Create a value of CreateVPCAssociationAuthorizationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateVPCAssociationAuthorizationResponse', createVPCAssociationAuthorizationResponse_httpStatus - The response's http status code.

$sel:hostedZoneId:CreateVPCAssociationAuthorization', createVPCAssociationAuthorizationResponse_hostedZoneId - The ID of the hosted zone that you authorized associating a VPC with.

$sel:vpc:CreateVPCAssociationAuthorization', createVPCAssociationAuthorizationResponse_vpc - The VPC that you authorized associating with a hosted zone.

Response Lenses

createVPCAssociationAuthorizationResponse_httpStatus :: Lens' CreateVPCAssociationAuthorizationResponse Int

The response's http status code.

createVPCAssociationAuthorizationResponse_hostedZoneId :: Lens' CreateVPCAssociationAuthorizationResponse ResourceId

The ID of the hosted zone that you authorized associating a VPC with.

createVPCAssociationAuthorizationResponse_vpc :: Lens' CreateVPCAssociationAuthorizationResponse VPC

The VPC that you authorized associating with a hosted zone.