Network.AWS.Route53.DeactivateKeySigningKey

Description

Deactivates a key-signing key (KSK) so that it will not be used for signing by DNSSEC. This operation changes the KSK status to INACTIVE.

Creating a Request

data DeactivateKeySigningKey

See: newDeactivateKeySigningKey smart constructor.

Constructors

DeactivateKeySigningKey'

Fields

hostedZoneId :: ResourceId

A unique string used to identify a hosted zone.

name :: Text

A string used to identify a key-signing key (KSK).

newDeactivateKeySigningKey

Arguments

:: ResourceId
$sel:hostedZoneId:DeactivateKeySigningKey'

-> Text
$sel:name:DeactivateKeySigningKey'

-> DeactivateKeySigningKey

Create a value of DeactivateKeySigningKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostedZoneId:DeactivateKeySigningKey', deactivateKeySigningKey_hostedZoneId - A unique string used to identify a hosted zone.

$sel:name:DeactivateKeySigningKey', deactivateKeySigningKey_name - A string used to identify a key-signing key (KSK).

Request Lenses

deactivateKeySigningKey_hostedZoneId :: Lens' DeactivateKeySigningKey ResourceId

A unique string used to identify a hosted zone.

deactivateKeySigningKey_name :: Lens' DeactivateKeySigningKey Text

A string used to identify a key-signing key (KSK).

Destructuring the Response

data DeactivateKeySigningKeyResponse

See: newDeactivateKeySigningKeyResponse smart constructor.

Constructors

DeactivateKeySigningKeyResponse'

Fields

httpStatus :: Int

The response's http status code.

changeInfo :: ChangeInfo

newDeactivateKeySigningKeyResponse

Arguments

:: Int
$sel:httpStatus:DeactivateKeySigningKeyResponse'

-> ChangeInfo
$sel:changeInfo:DeactivateKeySigningKeyResponse'

-> DeactivateKeySigningKeyResponse

Create a value of DeactivateKeySigningKeyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeactivateKeySigningKeyResponse', deactivateKeySigningKeyResponse_httpStatus - The response's http status code.

$sel:changeInfo:DeactivateKeySigningKeyResponse', deactivateKeySigningKeyResponse_changeInfo - Undocumented member.

Response Lenses

deactivateKeySigningKeyResponse_httpStatus :: Lens' DeactivateKeySigningKeyResponse Int

The response's http status code.

deactivateKeySigningKeyResponse_changeInfo :: Lens' DeactivateKeySigningKeyResponse ChangeInfo

Undocumented member.