Network.AWS.Route53.DeleteHealthCheck

Description

Deletes a health check.

Amazon Route 53 does not prevent you from deleting a health check even if the health check is associated with one or more resource record sets. If you delete a health check and you don't update the associated resource record sets, the future status of the health check can't be predicted and may change. This will affect the routing of DNS queries for your DNS failover configuration. For more information, see Replacing and Deleting Health Checks in the Amazon Route 53 Developer Guide.

If you're using Cloud Map and you configured Cloud Map to create a Route 53 health check when you register an instance, you can't use the Route 53 DeleteHealthCheck command to delete the health check. The health check is deleted automatically when you deregister the instance; there can be a delay of several hours before the health check is deleted from Route 53.

Creating a Request

data DeleteHealthCheck

This action deletes a health check.

See: newDeleteHealthCheck smart constructor.

Constructors

DeleteHealthCheck'

Fields

healthCheckId :: Text

The ID of the health check that you want to delete.

newDeleteHealthCheck

Arguments

:: Text
$sel:healthCheckId:DeleteHealthCheck'

-> DeleteHealthCheck

Create a value of DeleteHealthCheck with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:healthCheckId:DeleteHealthCheck', deleteHealthCheck_healthCheckId - The ID of the health check that you want to delete.

Request Lenses

deleteHealthCheck_healthCheckId :: Lens' DeleteHealthCheck Text

The ID of the health check that you want to delete.

Destructuring the Response

data DeleteHealthCheckResponse

An empty element.

See: newDeleteHealthCheckResponse smart constructor.

Constructors

DeleteHealthCheckResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteHealthCheckResponse

Arguments

:: Int
$sel:httpStatus:DeleteHealthCheckResponse'

-> DeleteHealthCheckResponse

Create a value of DeleteHealthCheckResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteHealthCheckResponse', deleteHealthCheckResponse_httpStatus - The response's http status code.

Response Lenses

deleteHealthCheckResponse_httpStatus :: Lens' DeleteHealthCheckResponse Int

The response's http status code.