Network.AWS.Route53.DisableHostedZoneDNSSEC

Description

Disables DNSSEC signing in a specific hosted zone. This action does not deactivate any key-signing keys (KSKs) that are active in the hosted zone.

Creating a Request

data DisableHostedZoneDNSSEC

See: newDisableHostedZoneDNSSEC smart constructor.

Constructors

DisableHostedZoneDNSSEC'

Fields

hostedZoneId :: ResourceId

A unique string used to identify a hosted zone.

newDisableHostedZoneDNSSEC

Arguments

:: ResourceId
$sel:hostedZoneId:DisableHostedZoneDNSSEC'

-> DisableHostedZoneDNSSEC

Create a value of DisableHostedZoneDNSSEC with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostedZoneId:DisableHostedZoneDNSSEC', disableHostedZoneDNSSEC_hostedZoneId - A unique string used to identify a hosted zone.

Request Lenses

disableHostedZoneDNSSEC_hostedZoneId :: Lens' DisableHostedZoneDNSSEC ResourceId

A unique string used to identify a hosted zone.

Destructuring the Response

data DisableHostedZoneDNSSECResponse

See: newDisableHostedZoneDNSSECResponse smart constructor.

Constructors

DisableHostedZoneDNSSECResponse'

Fields

httpStatus :: Int

The response's http status code.

changeInfo :: ChangeInfo

newDisableHostedZoneDNSSECResponse

Arguments

:: Int
$sel:httpStatus:DisableHostedZoneDNSSECResponse'

-> ChangeInfo
$sel:changeInfo:DisableHostedZoneDNSSECResponse'

-> DisableHostedZoneDNSSECResponse

Create a value of DisableHostedZoneDNSSECResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisableHostedZoneDNSSECResponse', disableHostedZoneDNSSECResponse_httpStatus - The response's http status code.

$sel:changeInfo:DisableHostedZoneDNSSECResponse', disableHostedZoneDNSSECResponse_changeInfo - Undocumented member.

Response Lenses

disableHostedZoneDNSSECResponse_httpStatus :: Lens' DisableHostedZoneDNSSECResponse Int

The response's http status code.

disableHostedZoneDNSSECResponse_changeInfo :: Lens' DisableHostedZoneDNSSECResponse ChangeInfo

Undocumented member.