Network.AWS.Route53.DisassociateVPCFromHostedZone

Description

Disassociates an Amazon Virtual Private Cloud (Amazon VPC) from an Amazon Route 53 private hosted zone. Note the following:

You can't disassociate the last Amazon VPC from a private hosted zone.

You can't convert a private hosted zone into a public hosted zone.

You can submit a DisassociateVPCFromHostedZone request using either the account that created the hosted zone or the account that created the Amazon VPC.

Some services, such as Cloud Map and Amazon Elastic File System (Amazon EFS) automatically create hosted zones and associate VPCs with the hosted zones. A service can create a hosted zone using your account or using its own account. You can disassociate a VPC from a hosted zone only if the service created the hosted zone using your account.

When you run DisassociateVPCFromHostedZone, if the hosted zone has a value for OwningAccount, you can use DisassociateVPCFromHostedZone. If the hosted zone has a value for OwningService, you can't use DisassociateVPCFromHostedZone.

Creating a Request

data DisassociateVPCFromHostedZone

A complex type that contains information about the VPC that you want to disassociate from a specified private hosted zone.

See: newDisassociateVPCFromHostedZone smart constructor.

Constructors

DisassociateVPCFromHostedZone'

Fields

comment :: Maybe Text

Optional: A comment about the disassociation request.

hostedZoneId :: ResourceId

The ID of the private hosted zone that you want to disassociate a VPC from.

vpc :: VPC

A complex type that contains information about the VPC that you're disassociating from the specified hosted zone.

newDisassociateVPCFromHostedZone

Arguments

:: ResourceId
$sel:hostedZoneId:DisassociateVPCFromHostedZone'

-> VPC
$sel:vpc:DisassociateVPCFromHostedZone'

-> DisassociateVPCFromHostedZone

Create a value of DisassociateVPCFromHostedZone with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:comment:DisassociateVPCFromHostedZone', disassociateVPCFromHostedZone_comment - Optional: A comment about the disassociation request.

$sel:hostedZoneId:DisassociateVPCFromHostedZone', disassociateVPCFromHostedZone_hostedZoneId - The ID of the private hosted zone that you want to disassociate a VPC from.

$sel:vpc:DisassociateVPCFromHostedZone', disassociateVPCFromHostedZone_vpc - A complex type that contains information about the VPC that you're disassociating from the specified hosted zone.

Request Lenses

disassociateVPCFromHostedZone_comment :: Lens' DisassociateVPCFromHostedZone (Maybe Text)

Optional: A comment about the disassociation request.

disassociateVPCFromHostedZone_hostedZoneId :: Lens' DisassociateVPCFromHostedZone ResourceId

The ID of the private hosted zone that you want to disassociate a VPC from.

disassociateVPCFromHostedZone_vpc :: Lens' DisassociateVPCFromHostedZone VPC

A complex type that contains information about the VPC that you're disassociating from the specified hosted zone.

Destructuring the Response

data DisassociateVPCFromHostedZoneResponse

A complex type that contains the response information for the disassociate request.

See: newDisassociateVPCFromHostedZoneResponse smart constructor.

Constructors

DisassociateVPCFromHostedZoneResponse'

Fields

httpStatus :: Int

The response's http status code.

changeInfo :: ChangeInfo

A complex type that describes the changes made to the specified private hosted zone.

newDisassociateVPCFromHostedZoneResponse

Arguments

:: Int
$sel:httpStatus:DisassociateVPCFromHostedZoneResponse'

-> ChangeInfo
$sel:changeInfo:DisassociateVPCFromHostedZoneResponse'

-> DisassociateVPCFromHostedZoneResponse

Create a value of DisassociateVPCFromHostedZoneResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisassociateVPCFromHostedZoneResponse', disassociateVPCFromHostedZoneResponse_httpStatus - The response's http status code.

$sel:changeInfo:DisassociateVPCFromHostedZoneResponse', disassociateVPCFromHostedZoneResponse_changeInfo - A complex type that describes the changes made to the specified private hosted zone.

Response Lenses

disassociateVPCFromHostedZoneResponse_httpStatus :: Lens' DisassociateVPCFromHostedZoneResponse Int

The response's http status code.

disassociateVPCFromHostedZoneResponse_changeInfo :: Lens' DisassociateVPCFromHostedZoneResponse ChangeInfo

A complex type that describes the changes made to the specified private hosted zone.