Network.AWS.Route53.GetAccountLimit

Description

Gets the specified limit for the current account, for example, the maximum number of health checks that you can create using the account.

For the default limit, see Limits in the Amazon Route 53 Developer Guide. To request a higher limit, open a case.

You can also view account limits in Amazon Web Services Trusted Advisor. Sign in to the Amazon Web Services Management Console and open the Trusted Advisor console at https://console.aws.amazon.com/trustedadvisor/. Then choose Service limits in the navigation pane.

Creating a Request

data GetAccountLimit

A complex type that contains information about the request to create a hosted zone.

See: newGetAccountLimit smart constructor.

Constructors

GetAccountLimit'

Fields

type' :: AccountLimitType

The limit that you want to get. Valid values include the following:

MAX_HEALTH_CHECKS_BY_OWNER: The maximum number of health checks that you can create using the current account.

MAX_HOSTED_ZONES_BY_OWNER: The maximum number of hosted zones that you can create using the current account.

MAX_REUSABLE_DELEGATION_SETS_BY_OWNER: The maximum number of reusable delegation sets that you can create using the current account.

MAX_TRAFFIC_POLICIES_BY_OWNER: The maximum number of traffic policies that you can create using the current account.

MAX_TRAFFIC_POLICY_INSTANCES_BY_OWNER: The maximum number of traffic policy instances that you can create using the current account. (Traffic policy instances are referred to as traffic flow policy records in the Amazon Route 53 console.)

newGetAccountLimit

Arguments

:: AccountLimitType
$sel:type':GetAccountLimit'

-> GetAccountLimit

Create a value of GetAccountLimit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':GetAccountLimit', getAccountLimit_type - The limit that you want to get. Valid values include the following:

MAX_HEALTH_CHECKS_BY_OWNER: The maximum number of health checks that you can create using the current account.

MAX_HOSTED_ZONES_BY_OWNER: The maximum number of hosted zones that you can create using the current account.

MAX_REUSABLE_DELEGATION_SETS_BY_OWNER: The maximum number of reusable delegation sets that you can create using the current account.

MAX_TRAFFIC_POLICIES_BY_OWNER: The maximum number of traffic policies that you can create using the current account.

MAX_TRAFFIC_POLICY_INSTANCES_BY_OWNER: The maximum number of traffic policy instances that you can create using the current account. (Traffic policy instances are referred to as traffic flow policy records in the Amazon Route 53 console.)

Request Lenses

getAccountLimit_type :: Lens' GetAccountLimit AccountLimitType

The limit that you want to get. Valid values include the following:

MAX_HEALTH_CHECKS_BY_OWNER: The maximum number of health checks that you can create using the current account.

MAX_HOSTED_ZONES_BY_OWNER: The maximum number of hosted zones that you can create using the current account.

MAX_REUSABLE_DELEGATION_SETS_BY_OWNER: The maximum number of reusable delegation sets that you can create using the current account.

MAX_TRAFFIC_POLICIES_BY_OWNER: The maximum number of traffic policies that you can create using the current account.

MAX_TRAFFIC_POLICY_INSTANCES_BY_OWNER: The maximum number of traffic policy instances that you can create using the current account. (Traffic policy instances are referred to as traffic flow policy records in the Amazon Route 53 console.)

Destructuring the Response

data GetAccountLimitResponse

A complex type that contains the requested limit.

See: newGetAccountLimitResponse smart constructor.

Constructors

GetAccountLimitResponse'

Fields

httpStatus :: Int

The response's http status code.

limit :: AccountLimit

The current setting for the specified limit. For example, if you specified MAX_HEALTH_CHECKS_BY_OWNER for the value of Type in the request, the value of Limit is the maximum number of health checks that you can create using the current account.

count :: Natural

The current number of entities that you have created of the specified type. For example, if you specified MAX_HEALTH_CHECKS_BY_OWNER for the value of Type in the request, the value of Count is the current number of health checks that you have created using the current account.

newGetAccountLimitResponse

Arguments

:: Int
$sel:httpStatus:GetAccountLimitResponse'

-> AccountLimit
$sel:limit:GetAccountLimitResponse'

-> Natural
$sel:count:GetAccountLimitResponse'

-> GetAccountLimitResponse

Create a value of GetAccountLimitResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetAccountLimitResponse', getAccountLimitResponse_httpStatus - The response's http status code.

$sel:limit:GetAccountLimitResponse', getAccountLimitResponse_limit - The current setting for the specified limit. For example, if you specified MAX_HEALTH_CHECKS_BY_OWNER for the value of Type in the request, the value of Limit is the maximum number of health checks that you can create using the current account.

$sel:count:GetAccountLimitResponse', getAccountLimitResponse_count - The current number of entities that you have created of the specified type. For example, if you specified MAX_HEALTH_CHECKS_BY_OWNER for the value of Type in the request, the value of Count is the current number of health checks that you have created using the current account.

Response Lenses

getAccountLimitResponse_httpStatus :: Lens' GetAccountLimitResponse Int

The response's http status code.

getAccountLimitResponse_limit :: Lens' GetAccountLimitResponse AccountLimit

The current setting for the specified limit. For example, if you specified MAX_HEALTH_CHECKS_BY_OWNER for the value of Type in the request, the value of Limit is the maximum number of health checks that you can create using the current account.

getAccountLimitResponse_count :: Lens' GetAccountLimitResponse Natural

The current number of entities that you have created of the specified type. For example, if you specified MAX_HEALTH_CHECKS_BY_OWNER for the value of Type in the request, the value of Count is the current number of health checks that you have created using the current account.