Network.AWS.Route53.GetDNSSEC

Description

Returns information about DNSSEC for a specific hosted zone, including the key-signing keys (KSKs) in the hosted zone.

Creating a Request

data GetDNSSEC

See: newGetDNSSEC smart constructor.

Constructors

GetDNSSEC'

Fields

hostedZoneId :: ResourceId

A unique string used to identify a hosted zone.

newGetDNSSEC

Arguments

:: ResourceId
$sel:hostedZoneId:GetDNSSEC'

-> GetDNSSEC

Create a value of GetDNSSEC with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostedZoneId:GetDNSSEC', getDNSSEC_hostedZoneId - A unique string used to identify a hosted zone.

Request Lenses

getDNSSEC_hostedZoneId :: Lens' GetDNSSEC ResourceId

A unique string used to identify a hosted zone.

Destructuring the Response

data GetDNSSECResponse

See: newGetDNSSECResponse smart constructor.

Constructors

GetDNSSECResponse'

Fields

httpStatus :: Int

The response's http status code.

status :: DNSSECStatus

A string repesenting the status of DNSSEC.

keySigningKeys :: [KeySigningKey]

The key-signing keys (KSKs) in your account.

newGetDNSSECResponse

Arguments

:: Int
$sel:httpStatus:GetDNSSECResponse'

-> DNSSECStatus
$sel:status:GetDNSSECResponse'

-> GetDNSSECResponse

Create a value of GetDNSSECResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetDNSSECResponse', getDNSSECResponse_httpStatus - The response's http status code.

$sel:status:GetDNSSECResponse', getDNSSECResponse_status - A string repesenting the status of DNSSEC.

$sel:keySigningKeys:GetDNSSECResponse', getDNSSECResponse_keySigningKeys - The key-signing keys (KSKs) in your account.

Response Lenses

getDNSSECResponse_httpStatus :: Lens' GetDNSSECResponse Int

The response's http status code.

getDNSSECResponse_status :: Lens' GetDNSSECResponse DNSSECStatus

A string repesenting the status of DNSSEC.

getDNSSECResponse_keySigningKeys :: Lens' GetDNSSECResponse [KeySigningKey]

The key-signing keys (KSKs) in your account.