Network.AWS.Route53.GetHealthCheck

Description

Gets information about a specified health check.

Creating a Request

data GetHealthCheck

A request to get information about a specified health check.

See: newGetHealthCheck smart constructor.

Constructors

GetHealthCheck'

Fields

healthCheckId :: Text

The identifier that Amazon Route 53 assigned to the health check when you created it. When you add or update a resource record set, you use this value to specify which health check to use. The value can be up to 64 characters long.

newGetHealthCheck

Arguments

:: Text
$sel:healthCheckId:GetHealthCheck'

-> GetHealthCheck

Create a value of GetHealthCheck with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:healthCheckId:GetHealthCheck', getHealthCheck_healthCheckId - The identifier that Amazon Route 53 assigned to the health check when you created it. When you add or update a resource record set, you use this value to specify which health check to use. The value can be up to 64 characters long.

Request Lenses

getHealthCheck_healthCheckId :: Lens' GetHealthCheck Text

The identifier that Amazon Route 53 assigned to the health check when you created it. When you add or update a resource record set, you use this value to specify which health check to use. The value can be up to 64 characters long.

Destructuring the Response

data GetHealthCheckResponse

A complex type that contains the response to a GetHealthCheck request.

See: newGetHealthCheckResponse smart constructor.

Constructors

GetHealthCheckResponse'

Fields

httpStatus :: Int

The response's http status code.

healthCheck :: HealthCheck

A complex type that contains information about one health check that is associated with the current Amazon Web Services account.

newGetHealthCheckResponse

Arguments

:: Int
$sel:httpStatus:GetHealthCheckResponse'

-> HealthCheck
$sel:healthCheck:GetHealthCheckResponse'

-> GetHealthCheckResponse

Create a value of GetHealthCheckResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetHealthCheckResponse', getHealthCheckResponse_httpStatus - The response's http status code.

$sel:healthCheck:GetHealthCheckResponse', getHealthCheckResponse_healthCheck - A complex type that contains information about one health check that is associated with the current Amazon Web Services account.

Response Lenses

getHealthCheckResponse_httpStatus :: Lens' GetHealthCheckResponse Int

The response's http status code.

getHealthCheckResponse_healthCheck :: Lens' GetHealthCheckResponse HealthCheck

A complex type that contains information about one health check that is associated with the current Amazon Web Services account.