Network.AWS.Route53.GetHealthCheckCount

Description

Retrieves the number of health checks that are associated with the current Amazon Web Services account.

Creating a Request

data GetHealthCheckCount

A request for the number of health checks that are associated with the current Amazon Web Services account.

See: newGetHealthCheckCount smart constructor.

Constructors

GetHealthCheckCount'

newGetHealthCheckCount :: GetHealthCheckCount

Create a value of GetHealthCheckCount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetHealthCheckCountResponse

A complex type that contains the response to a GetHealthCheckCount request.

See: newGetHealthCheckCountResponse smart constructor.

Constructors

GetHealthCheckCountResponse'

Fields

httpStatus :: Int

The response's http status code.

healthCheckCount :: Integer

The number of health checks associated with the current Amazon Web Services account.

newGetHealthCheckCountResponse

Arguments

:: Int
$sel:httpStatus:GetHealthCheckCountResponse'

-> Integer
$sel:healthCheckCount:GetHealthCheckCountResponse'

-> GetHealthCheckCountResponse

Create a value of GetHealthCheckCountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetHealthCheckCountResponse', getHealthCheckCountResponse_httpStatus - The response's http status code.

$sel:healthCheckCount:GetHealthCheckCountResponse', getHealthCheckCountResponse_healthCheckCount - The number of health checks associated with the current Amazon Web Services account.

Response Lenses

getHealthCheckCountResponse_httpStatus :: Lens' GetHealthCheckCountResponse Int

The response's http status code.

getHealthCheckCountResponse_healthCheckCount :: Lens' GetHealthCheckCountResponse Integer

The number of health checks associated with the current Amazon Web Services account.