Network.AWS.Route53.ListGeoLocations

Description

Retrieves a list of supported geographic locations.

Countries are listed first, and continents are listed last. If Amazon Route 53 supports subdivisions for a country (for example, states or provinces), the subdivisions for that country are listed in alphabetical order immediately after the corresponding country.

Route 53 does not perform authorization for this API because it retrieves information that is already available to the public.

For a list of supported geolocation codes, see the GeoLocation data type.

Creating a Request

data ListGeoLocations

A request to get a list of geographic locations that Amazon Route 53 supports for geolocation resource record sets.

See: newListGeoLocations smart constructor.

Constructors

ListGeoLocations'

Fields

startSubdivisionCode :: Maybe Text

The code for the state of the United States with which you want to start listing locations that Amazon Route 53 supports for geolocation. If Route 53 has already returned a page or more of results, if IsTruncated is true, and if NextSubdivisionCode from the previous response has a value, enter that value in startsubdivisioncode to return the next page of results.

To list subdivisions (U.S. states), you must include both startcountrycode and startsubdivisioncode.

startCountryCode :: Maybe Text

The code for the country with which you want to start listing locations that Amazon Route 53 supports for geolocation. If Route 53 has already returned a page or more of results, if IsTruncated is true, and if NextCountryCode from the previous response has a value, enter that value in startcountrycode to return the next page of results.

startContinentCode :: Maybe Text

The code for the continent with which you want to start listing locations that Amazon Route 53 supports for geolocation. If Route 53 has already returned a page or more of results, if IsTruncated is true, and if NextContinentCode from the previous response has a value, enter that value in startcontinentcode to return the next page of results.

Include startcontinentcode only if you want to list continents. Don't include startcontinentcode when you're listing countries or countries with their subdivisions.

maxItems :: Maybe Text

(Optional) The maximum number of geolocations to be included in the response body for this request. If more than maxitems geolocations remain to be listed, then the value of the IsTruncated element in the response is true.

newListGeoLocations :: ListGeoLocations

Create a value of ListGeoLocations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startSubdivisionCode:ListGeoLocations', listGeoLocations_startSubdivisionCode - The code for the state of the United States with which you want to start listing locations that Amazon Route 53 supports for geolocation. If Route 53 has already returned a page or more of results, if IsTruncated is true, and if NextSubdivisionCode from the previous response has a value, enter that value in startsubdivisioncode to return the next page of results.

To list subdivisions (U.S. states), you must include both startcountrycode and startsubdivisioncode.

$sel:startCountryCode:ListGeoLocations', listGeoLocations_startCountryCode - The code for the country with which you want to start listing locations that Amazon Route 53 supports for geolocation. If Route 53 has already returned a page or more of results, if IsTruncated is true, and if NextCountryCode from the previous response has a value, enter that value in startcountrycode to return the next page of results.

$sel:startContinentCode:ListGeoLocations', listGeoLocations_startContinentCode - The code for the continent with which you want to start listing locations that Amazon Route 53 supports for geolocation. If Route 53 has already returned a page or more of results, if IsTruncated is true, and if NextContinentCode from the previous response has a value, enter that value in startcontinentcode to return the next page of results.

Include startcontinentcode only if you want to list continents. Don't include startcontinentcode when you're listing countries or countries with their subdivisions.

$sel:maxItems:ListGeoLocations', listGeoLocations_maxItems - (Optional) The maximum number of geolocations to be included in the response body for this request. If more than maxitems geolocations remain to be listed, then the value of the IsTruncated element in the response is true.

Request Lenses

listGeoLocations_startSubdivisionCode :: Lens' ListGeoLocations (Maybe Text)

The code for the state of the United States with which you want to start listing locations that Amazon Route 53 supports for geolocation. If Route 53 has already returned a page or more of results, if IsTruncated is true, and if NextSubdivisionCode from the previous response has a value, enter that value in startsubdivisioncode to return the next page of results.

To list subdivisions (U.S. states), you must include both startcountrycode and startsubdivisioncode.

listGeoLocations_startCountryCode :: Lens' ListGeoLocations (Maybe Text)

The code for the country with which you want to start listing locations that Amazon Route 53 supports for geolocation. If Route 53 has already returned a page or more of results, if IsTruncated is true, and if NextCountryCode from the previous response has a value, enter that value in startcountrycode to return the next page of results.

listGeoLocations_startContinentCode :: Lens' ListGeoLocations (Maybe Text)

The code for the continent with which you want to start listing locations that Amazon Route 53 supports for geolocation. If Route 53 has already returned a page or more of results, if IsTruncated is true, and if NextContinentCode from the previous response has a value, enter that value in startcontinentcode to return the next page of results.

Include startcontinentcode only if you want to list continents. Don't include startcontinentcode when you're listing countries or countries with their subdivisions.

listGeoLocations_maxItems :: Lens' ListGeoLocations (Maybe Text)

(Optional) The maximum number of geolocations to be included in the response body for this request. If more than maxitems geolocations remain to be listed, then the value of the IsTruncated element in the response is true.

Destructuring the Response

data ListGeoLocationsResponse

A complex type containing the response information for the request.

See: newListGeoLocationsResponse smart constructor.

Constructors

ListGeoLocationsResponse'

Fields

nextSubdivisionCode :: Maybe Text

If IsTruncated is true, you can make a follow-up request to display more locations. Enter the value of NextSubdivisionCode in the startsubdivisioncode parameter in another ListGeoLocations request.

nextContinentCode :: Maybe Text

If IsTruncated is true, you can make a follow-up request to display more locations. Enter the value of NextContinentCode in the startcontinentcode parameter in another ListGeoLocations request.

nextCountryCode :: Maybe Text

If IsTruncated is true, you can make a follow-up request to display more locations. Enter the value of NextCountryCode in the startcountrycode parameter in another ListGeoLocations request.

httpStatus :: Int

The response's http status code.

geoLocationDetailsList :: [GeoLocationDetails]

A complex type that contains one GeoLocationDetails element for each location that Amazon Route 53 supports for geolocation.

isTruncated :: Bool

A value that indicates whether more locations remain to be listed after the last location in this response. If so, the value of IsTruncated is true. To get more values, submit another request and include the values of NextContinentCode, NextCountryCode, and NextSubdivisionCode in the startcontinentcode, startcountrycode, and startsubdivisioncode, as applicable.

maxItems :: Text

The value that you specified for MaxItems in the request.

newListGeoLocationsResponse

Arguments

:: Int
$sel:httpStatus:ListGeoLocationsResponse'

-> Bool
$sel:isTruncated:ListGeoLocationsResponse'

-> Text
$sel:maxItems:ListGeoLocations'

-> ListGeoLocationsResponse

Create a value of ListGeoLocationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextSubdivisionCode:ListGeoLocationsResponse', listGeoLocationsResponse_nextSubdivisionCode - If IsTruncated is true, you can make a follow-up request to display more locations. Enter the value of NextSubdivisionCode in the startsubdivisioncode parameter in another ListGeoLocations request.

$sel:nextContinentCode:ListGeoLocationsResponse', listGeoLocationsResponse_nextContinentCode - If IsTruncated is true, you can make a follow-up request to display more locations. Enter the value of NextContinentCode in the startcontinentcode parameter in another ListGeoLocations request.

$sel:nextCountryCode:ListGeoLocationsResponse', listGeoLocationsResponse_nextCountryCode - If IsTruncated is true, you can make a follow-up request to display more locations. Enter the value of NextCountryCode in the startcountrycode parameter in another ListGeoLocations request.

$sel:httpStatus:ListGeoLocationsResponse', listGeoLocationsResponse_httpStatus - The response's http status code.

$sel:geoLocationDetailsList:ListGeoLocationsResponse', listGeoLocationsResponse_geoLocationDetailsList - A complex type that contains one GeoLocationDetails element for each location that Amazon Route 53 supports for geolocation.

$sel:isTruncated:ListGeoLocationsResponse', listGeoLocationsResponse_isTruncated - A value that indicates whether more locations remain to be listed after the last location in this response. If so, the value of IsTruncated is true. To get more values, submit another request and include the values of NextContinentCode, NextCountryCode, and NextSubdivisionCode in the startcontinentcode, startcountrycode, and startsubdivisioncode, as applicable.

$sel:maxItems:ListGeoLocations', listGeoLocationsResponse_maxItems - The value that you specified for MaxItems in the request.

Response Lenses

listGeoLocationsResponse_nextSubdivisionCode :: Lens' ListGeoLocationsResponse (Maybe Text)

If IsTruncated is true, you can make a follow-up request to display more locations. Enter the value of NextSubdivisionCode in the startsubdivisioncode parameter in another ListGeoLocations request.

listGeoLocationsResponse_nextContinentCode :: Lens' ListGeoLocationsResponse (Maybe Text)

If IsTruncated is true, you can make a follow-up request to display more locations. Enter the value of NextContinentCode in the startcontinentcode parameter in another ListGeoLocations request.

listGeoLocationsResponse_nextCountryCode :: Lens' ListGeoLocationsResponse (Maybe Text)

If IsTruncated is true, you can make a follow-up request to display more locations. Enter the value of NextCountryCode in the startcountrycode parameter in another ListGeoLocations request.

listGeoLocationsResponse_httpStatus :: Lens' ListGeoLocationsResponse Int

The response's http status code.

listGeoLocationsResponse_geoLocationDetailsList :: Lens' ListGeoLocationsResponse [GeoLocationDetails]

A complex type that contains one GeoLocationDetails element for each location that Amazon Route 53 supports for geolocation.

listGeoLocationsResponse_isTruncated :: Lens' ListGeoLocationsResponse Bool

A value that indicates whether more locations remain to be listed after the last location in this response. If so, the value of IsTruncated is true. To get more values, submit another request and include the values of NextContinentCode, NextCountryCode, and NextSubdivisionCode in the startcontinentcode, startcountrycode, and startsubdivisioncode, as applicable.

listGeoLocationsResponse_maxItems :: Lens' ListGeoLocationsResponse Text

The value that you specified for MaxItems in the request.