Network.AWS.Route53.ListHealthChecks

Description

Retrieve a list of the health checks that are associated with the current Amazon Web Services account.

This operation returns paginated results.

Creating a Request

data ListHealthChecks

A request to retrieve a list of the health checks that are associated with the current Amazon Web Services account.

See: newListHealthChecks smart constructor.

Constructors

ListHealthChecks'

Fields

marker :: Maybe Text

If the value of IsTruncated in the previous response was true, you have more health checks. To get another group, submit another ListHealthChecks request.

For the value of marker, specify the value of NextMarker from the previous response, which is the ID of the first health check that Amazon Route 53 will return if you submit another request.

If the value of IsTruncated in the previous response was false, there are no more health checks to get.

maxItems :: Maybe Text

The maximum number of health checks that you want ListHealthChecks to return in response to the current request. Amazon Route 53 returns a maximum of 100 items. If you set MaxItems to a value greater than 100, Route 53 returns only the first 100 health checks.

newListHealthChecks :: ListHealthChecks

Create a value of ListHealthChecks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListHealthChecks', listHealthChecks_marker - If the value of IsTruncated in the previous response was true, you have more health checks. To get another group, submit another ListHealthChecks request.

For the value of marker, specify the value of NextMarker from the previous response, which is the ID of the first health check that Amazon Route 53 will return if you submit another request.

If the value of IsTruncated in the previous response was false, there are no more health checks to get.

$sel:maxItems:ListHealthChecks', listHealthChecks_maxItems - The maximum number of health checks that you want ListHealthChecks to return in response to the current request. Amazon Route 53 returns a maximum of 100 items. If you set MaxItems to a value greater than 100, Route 53 returns only the first 100 health checks.

Request Lenses

listHealthChecks_marker :: Lens' ListHealthChecks (Maybe Text)

If the value of IsTruncated in the previous response was true, you have more health checks. To get another group, submit another ListHealthChecks request.

For the value of marker, specify the value of NextMarker from the previous response, which is the ID of the first health check that Amazon Route 53 will return if you submit another request.

If the value of IsTruncated in the previous response was false, there are no more health checks to get.

listHealthChecks_maxItems :: Lens' ListHealthChecks (Maybe Text)

The maximum number of health checks that you want ListHealthChecks to return in response to the current request. Amazon Route 53 returns a maximum of 100 items. If you set MaxItems to a value greater than 100, Route 53 returns only the first 100 health checks.

Destructuring the Response

data ListHealthChecksResponse

A complex type that contains the response to a ListHealthChecks request.

See: newListHealthChecksResponse smart constructor.

Constructors

ListHealthChecksResponse'

Fields

nextMarker :: Maybe Text

If IsTruncated is true, the value of NextMarker identifies the first health check that Amazon Route 53 returns if you submit another ListHealthChecks request and specify the value of NextMarker in the marker parameter.

httpStatus :: Int

The response's http status code.

healthChecks :: [HealthCheck]

A complex type that contains one HealthCheck element for each health check that is associated with the current Amazon Web Services account.

marker :: Text

For the second and subsequent calls to ListHealthChecks, Marker is the value that you specified for the marker parameter in the previous request.

isTruncated :: Bool

A flag that indicates whether there are more health checks to be listed. If the response was truncated, you can get the next group of health checks by submitting another ListHealthChecks request and specifying the value of NextMarker in the marker parameter.

maxItems :: Text

The value that you specified for the maxitems parameter in the call to ListHealthChecks that produced the current response.

newListHealthChecksResponse

Arguments

:: Int
$sel:httpStatus:ListHealthChecksResponse'

-> Text
$sel:marker:ListHealthChecks'

-> Bool
$sel:isTruncated:ListHealthChecksResponse'

-> Text
$sel:maxItems:ListHealthChecks'

-> ListHealthChecksResponse

Create a value of ListHealthChecksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextMarker:ListHealthChecksResponse', listHealthChecksResponse_nextMarker - If IsTruncated is true, the value of NextMarker identifies the first health check that Amazon Route 53 returns if you submit another ListHealthChecks request and specify the value of NextMarker in the marker parameter.

$sel:httpStatus:ListHealthChecksResponse', listHealthChecksResponse_httpStatus - The response's http status code.

$sel:healthChecks:ListHealthChecksResponse', listHealthChecksResponse_healthChecks - A complex type that contains one HealthCheck element for each health check that is associated with the current Amazon Web Services account.

$sel:marker:ListHealthChecks', listHealthChecksResponse_marker - For the second and subsequent calls to ListHealthChecks, Marker is the value that you specified for the marker parameter in the previous request.

$sel:isTruncated:ListHealthChecksResponse', listHealthChecksResponse_isTruncated - A flag that indicates whether there are more health checks to be listed. If the response was truncated, you can get the next group of health checks by submitting another ListHealthChecks request and specifying the value of NextMarker in the marker parameter.

$sel:maxItems:ListHealthChecks', listHealthChecksResponse_maxItems - The value that you specified for the maxitems parameter in the call to ListHealthChecks that produced the current response.

Response Lenses

listHealthChecksResponse_nextMarker :: Lens' ListHealthChecksResponse (Maybe Text)

If IsTruncated is true, the value of NextMarker identifies the first health check that Amazon Route 53 returns if you submit another ListHealthChecks request and specify the value of NextMarker in the marker parameter.

listHealthChecksResponse_httpStatus :: Lens' ListHealthChecksResponse Int

The response's http status code.

listHealthChecksResponse_healthChecks :: Lens' ListHealthChecksResponse [HealthCheck]

A complex type that contains one HealthCheck element for each health check that is associated with the current Amazon Web Services account.

listHealthChecksResponse_marker :: Lens' ListHealthChecksResponse Text

For the second and subsequent calls to ListHealthChecks, Marker is the value that you specified for the marker parameter in the previous request.

listHealthChecksResponse_isTruncated :: Lens' ListHealthChecksResponse Bool

A flag that indicates whether there are more health checks to be listed. If the response was truncated, you can get the next group of health checks by submitting another ListHealthChecks request and specifying the value of NextMarker in the marker parameter.

listHealthChecksResponse_maxItems :: Lens' ListHealthChecksResponse Text

The value that you specified for the maxitems parameter in the call to ListHealthChecks that produced the current response.