Network.AWS.Route53.ListHostedZones

Description

Retrieves a list of the public and private hosted zones that are associated with the current Amazon Web Services account. The response includes a HostedZones child element for each hosted zone.

Amazon Route 53 returns a maximum of 100 items in each response. If you have a lot of hosted zones, you can use the maxitems parameter to list them in groups of up to 100.

This operation returns paginated results.

Creating a Request

data ListHostedZones

A request to retrieve a list of the public and private hosted zones that are associated with the current Amazon Web Services account.

See: newListHostedZones smart constructor.

Constructors

ListHostedZones'

Fields

delegationSetId :: Maybe ResourceId

If you're using reusable delegation sets and you want to list all of the hosted zones that are associated with a reusable delegation set, specify the ID of that reusable delegation set.

marker :: Maybe Text

If the value of IsTruncated in the previous response was true, you have more hosted zones. To get more hosted zones, submit another ListHostedZones request.

For the value of marker, specify the value of NextMarker from the previous response, which is the ID of the first hosted zone that Amazon Route 53 will return if you submit another request.

If the value of IsTruncated in the previous response was false, there are no more hosted zones to get.

maxItems :: Maybe Text

(Optional) The maximum number of hosted zones that you want Amazon Route 53 to return. If you have more than maxitems hosted zones, the value of IsTruncated in the response is true, and the value of NextMarker is the hosted zone ID of the first hosted zone that Route 53 will return if you submit another request.

newListHostedZones :: ListHostedZones

Create a value of ListHostedZones with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:delegationSetId:ListHostedZones', listHostedZones_delegationSetId - If you're using reusable delegation sets and you want to list all of the hosted zones that are associated with a reusable delegation set, specify the ID of that reusable delegation set.

$sel:marker:ListHostedZones', listHostedZones_marker - If the value of IsTruncated in the previous response was true, you have more hosted zones. To get more hosted zones, submit another ListHostedZones request.

For the value of marker, specify the value of NextMarker from the previous response, which is the ID of the first hosted zone that Amazon Route 53 will return if you submit another request.

If the value of IsTruncated in the previous response was false, there are no more hosted zones to get.

$sel:maxItems:ListHostedZones', listHostedZones_maxItems - (Optional) The maximum number of hosted zones that you want Amazon Route 53 to return. If you have more than maxitems hosted zones, the value of IsTruncated in the response is true, and the value of NextMarker is the hosted zone ID of the first hosted zone that Route 53 will return if you submit another request.

Request Lenses

listHostedZones_delegationSetId :: Lens' ListHostedZones (Maybe ResourceId)

If you're using reusable delegation sets and you want to list all of the hosted zones that are associated with a reusable delegation set, specify the ID of that reusable delegation set.

listHostedZones_marker :: Lens' ListHostedZones (Maybe Text)

If the value of IsTruncated in the previous response was true, you have more hosted zones. To get more hosted zones, submit another ListHostedZones request.

For the value of marker, specify the value of NextMarker from the previous response, which is the ID of the first hosted zone that Amazon Route 53 will return if you submit another request.

If the value of IsTruncated in the previous response was false, there are no more hosted zones to get.

listHostedZones_maxItems :: Lens' ListHostedZones (Maybe Text)

(Optional) The maximum number of hosted zones that you want Amazon Route 53 to return. If you have more than maxitems hosted zones, the value of IsTruncated in the response is true, and the value of NextMarker is the hosted zone ID of the first hosted zone that Route 53 will return if you submit another request.

Destructuring the Response

data ListHostedZonesResponse

See: newListHostedZonesResponse smart constructor.

Constructors

ListHostedZonesResponse'

Fields

marker :: Maybe Text

For the second and subsequent calls to ListHostedZones, Marker is the value that you specified for the marker parameter in the request that produced the current response.

nextMarker :: Maybe Text

If IsTruncated is true, the value of NextMarker identifies the first hosted zone in the next group of hosted zones. Submit another ListHostedZones request, and specify the value of NextMarker from the response in the marker parameter.

This element is present only if IsTruncated is true.

httpStatus :: Int

The response's http status code.

hostedZones :: [HostedZone]

A complex type that contains general information about the hosted zone.

isTruncated :: Bool

A flag indicating whether there are more hosted zones to be listed. If the response was truncated, you can get more hosted zones by submitting another ListHostedZones request and specifying the value of NextMarker in the marker parameter.

maxItems :: Text

The value that you specified for the maxitems parameter in the call to ListHostedZones that produced the current response.

newListHostedZonesResponse

Arguments

:: Int
$sel:httpStatus:ListHostedZonesResponse'

-> Bool
$sel:isTruncated:ListHostedZonesResponse'

-> Text
$sel:maxItems:ListHostedZones'

-> ListHostedZonesResponse

Create a value of ListHostedZonesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListHostedZones', listHostedZonesResponse_marker - For the second and subsequent calls to ListHostedZones, Marker is the value that you specified for the marker parameter in the request that produced the current response.

$sel:nextMarker:ListHostedZonesResponse', listHostedZonesResponse_nextMarker - If IsTruncated is true, the value of NextMarker identifies the first hosted zone in the next group of hosted zones. Submit another ListHostedZones request, and specify the value of NextMarker from the response in the marker parameter.

This element is present only if IsTruncated is true.

$sel:httpStatus:ListHostedZonesResponse', listHostedZonesResponse_httpStatus - The response's http status code.

$sel:hostedZones:ListHostedZonesResponse', listHostedZonesResponse_hostedZones - A complex type that contains general information about the hosted zone.

$sel:isTruncated:ListHostedZonesResponse', listHostedZonesResponse_isTruncated - A flag indicating whether there are more hosted zones to be listed. If the response was truncated, you can get more hosted zones by submitting another ListHostedZones request and specifying the value of NextMarker in the marker parameter.

$sel:maxItems:ListHostedZones', listHostedZonesResponse_maxItems - The value that you specified for the maxitems parameter in the call to ListHostedZones that produced the current response.

Response Lenses

listHostedZonesResponse_marker :: Lens' ListHostedZonesResponse (Maybe Text)

For the second and subsequent calls to ListHostedZones, Marker is the value that you specified for the marker parameter in the request that produced the current response.

listHostedZonesResponse_nextMarker :: Lens' ListHostedZonesResponse (Maybe Text)

If IsTruncated is true, the value of NextMarker identifies the first hosted zone in the next group of hosted zones. Submit another ListHostedZones request, and specify the value of NextMarker from the response in the marker parameter.

This element is present only if IsTruncated is true.

listHostedZonesResponse_httpStatus :: Lens' ListHostedZonesResponse Int

The response's http status code.

listHostedZonesResponse_hostedZones :: Lens' ListHostedZonesResponse [HostedZone]

A complex type that contains general information about the hosted zone.

listHostedZonesResponse_isTruncated :: Lens' ListHostedZonesResponse Bool

A flag indicating whether there are more hosted zones to be listed. If the response was truncated, you can get more hosted zones by submitting another ListHostedZones request and specifying the value of NextMarker in the marker parameter.

listHostedZonesResponse_maxItems :: Lens' ListHostedZonesResponse Text

The value that you specified for the maxitems parameter in the call to ListHostedZones that produced the current response.