Network.AWS.Route53.ListHostedZonesByVPC

Description

Lists all the private hosted zones that a specified VPC is associated with, regardless of which Amazon Web Services account or Amazon Web Services service owns the hosted zones. The HostedZoneOwner structure in the response contains one of the following values:

An OwningAccount element, which contains the account number of either the current Amazon Web Services account or another Amazon Web Services account. Some services, such as Cloud Map, create hosted zones using the current account.

An OwningService element, which identifies the Amazon Web Services service that created and owns the hosted zone. For example, if a hosted zone was created by Amazon Elastic File System (Amazon EFS), the value of Owner is efs.amazonaws.com.

Creating a Request

data ListHostedZonesByVPC

Lists all the private hosted zones that a specified VPC is associated with, regardless of which Amazon Web Services account created the hosted zones.

See: newListHostedZonesByVPC smart constructor.

Constructors

ListHostedZonesByVPC'

Fields

nextToken :: Maybe Text

If the previous response included a NextToken element, the specified VPC is associated with more hosted zones. To get more hosted zones, submit another ListHostedZonesByVPC request.

For the value of NextToken, specify the value of NextToken from the previous response.

If the previous response didn't include a NextToken element, there are no more hosted zones to get.

maxItems :: Maybe Text

(Optional) The maximum number of hosted zones that you want Amazon Route 53 to return. If the specified VPC is associated with more than MaxItems hosted zones, the response includes a NextToken element. NextToken contains an encrypted token that identifies the first hosted zone that Route 53 will return if you submit another request.

vPCId :: Text

The ID of the Amazon VPC that you want to list hosted zones for.

vPCRegion :: VPCRegion

For the Amazon VPC that you specified for VPCId, the Amazon Web Services Region that you created the VPC in.

newListHostedZonesByVPC

Arguments

:: Text
$sel:vPCId:ListHostedZonesByVPC'

-> VPCRegion
$sel:vPCRegion:ListHostedZonesByVPC'

-> ListHostedZonesByVPC

Create a value of ListHostedZonesByVPC with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListHostedZonesByVPC', listHostedZonesByVPC_nextToken - If the previous response included a NextToken element, the specified VPC is associated with more hosted zones. To get more hosted zones, submit another ListHostedZonesByVPC request.

For the value of NextToken, specify the value of NextToken from the previous response.

If the previous response didn't include a NextToken element, there are no more hosted zones to get.

$sel:maxItems:ListHostedZonesByVPC', listHostedZonesByVPC_maxItems - (Optional) The maximum number of hosted zones that you want Amazon Route 53 to return. If the specified VPC is associated with more than MaxItems hosted zones, the response includes a NextToken element. NextToken contains an encrypted token that identifies the first hosted zone that Route 53 will return if you submit another request.

$sel:vPCId:ListHostedZonesByVPC', listHostedZonesByVPC_vPCId - The ID of the Amazon VPC that you want to list hosted zones for.

$sel:vPCRegion:ListHostedZonesByVPC', listHostedZonesByVPC_vPCRegion - For the Amazon VPC that you specified for VPCId, the Amazon Web Services Region that you created the VPC in.

Request Lenses

listHostedZonesByVPC_nextToken :: Lens' ListHostedZonesByVPC (Maybe Text)

If the previous response included a NextToken element, the specified VPC is associated with more hosted zones. To get more hosted zones, submit another ListHostedZonesByVPC request.

For the value of NextToken, specify the value of NextToken from the previous response.

If the previous response didn't include a NextToken element, there are no more hosted zones to get.

listHostedZonesByVPC_maxItems :: Lens' ListHostedZonesByVPC (Maybe Text)

(Optional) The maximum number of hosted zones that you want Amazon Route 53 to return. If the specified VPC is associated with more than MaxItems hosted zones, the response includes a NextToken element. NextToken contains an encrypted token that identifies the first hosted zone that Route 53 will return if you submit another request.

listHostedZonesByVPC_vPCId :: Lens' ListHostedZonesByVPC Text

The ID of the Amazon VPC that you want to list hosted zones for.

listHostedZonesByVPC_vPCRegion :: Lens' ListHostedZonesByVPC VPCRegion

For the Amazon VPC that you specified for VPCId, the Amazon Web Services Region that you created the VPC in.

Destructuring the Response

data ListHostedZonesByVPCResponse

See: newListHostedZonesByVPCResponse smart constructor.

Constructors

ListHostedZonesByVPCResponse'

Fields

nextToken :: Maybe Text

The value that you will use for NextToken in the next ListHostedZonesByVPC request.

httpStatus :: Int

The response's http status code.

hostedZoneSummaries :: [HostedZoneSummary]

A list that contains one HostedZoneSummary element for each hosted zone that the specified Amazon VPC is associated with. Each HostedZoneSummary element contains the hosted zone name and ID, and information about who owns the hosted zone.

maxItems :: Text

The value that you specified for MaxItems in the most recent ListHostedZonesByVPC request.

newListHostedZonesByVPCResponse

Arguments

:: Int
$sel:httpStatus:ListHostedZonesByVPCResponse'

-> Text
$sel:maxItems:ListHostedZonesByVPC'

-> ListHostedZonesByVPCResponse

Create a value of ListHostedZonesByVPCResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListHostedZonesByVPC', listHostedZonesByVPCResponse_nextToken - The value that you will use for NextToken in the next ListHostedZonesByVPC request.

$sel:httpStatus:ListHostedZonesByVPCResponse', listHostedZonesByVPCResponse_httpStatus - The response's http status code.

$sel:hostedZoneSummaries:ListHostedZonesByVPCResponse', listHostedZonesByVPCResponse_hostedZoneSummaries - A list that contains one HostedZoneSummary element for each hosted zone that the specified Amazon VPC is associated with. Each HostedZoneSummary element contains the hosted zone name and ID, and information about who owns the hosted zone.

$sel:maxItems:ListHostedZonesByVPC', listHostedZonesByVPCResponse_maxItems - The value that you specified for MaxItems in the most recent ListHostedZonesByVPC request.

Response Lenses

listHostedZonesByVPCResponse_nextToken :: Lens' ListHostedZonesByVPCResponse (Maybe Text)

The value that you will use for NextToken in the next ListHostedZonesByVPC request.

listHostedZonesByVPCResponse_httpStatus :: Lens' ListHostedZonesByVPCResponse Int

The response's http status code.

listHostedZonesByVPCResponse_hostedZoneSummaries :: Lens' ListHostedZonesByVPCResponse [HostedZoneSummary]

A list that contains one HostedZoneSummary element for each hosted zone that the specified Amazon VPC is associated with. Each HostedZoneSummary element contains the hosted zone name and ID, and information about who owns the hosted zone.

listHostedZonesByVPCResponse_maxItems :: Lens' ListHostedZonesByVPCResponse Text

The value that you specified for MaxItems in the most recent ListHostedZonesByVPC request.