Network.AWS.Route53.ListQueryLoggingConfigs

Description

Lists the configurations for DNS query logging that are associated with the current Amazon Web Services account or the configuration that is associated with a specified hosted zone.

For more information about DNS query logs, see CreateQueryLoggingConfig. Additional information, including the format of DNS query logs, appears in Logging DNS Queries in the Amazon Route 53 Developer Guide.

This operation returns paginated results.

Creating a Request

data ListQueryLoggingConfigs

See: newListQueryLoggingConfigs smart constructor.

Constructors

ListQueryLoggingConfigs'

Fields

hostedZoneId :: Maybe ResourceId

(Optional) If you want to list the query logging configuration that is associated with a hosted zone, specify the ID in HostedZoneId.

If you don't specify a hosted zone ID, ListQueryLoggingConfigs returns all of the configurations that are associated with the current Amazon Web Services account.

nextToken :: Maybe Text

(Optional) If the current Amazon Web Services account has more than MaxResults query logging configurations, use NextToken to get the second and subsequent pages of results.

For the first ListQueryLoggingConfigs request, omit this value.

For the second and subsequent requests, get the value of NextToken from the previous response and specify that value for NextToken in the request.

maxResults :: Maybe Text

(Optional) The maximum number of query logging configurations that you want Amazon Route 53 to return in response to the current request. If the current Amazon Web Services account has more than MaxResults configurations, use the value of NextToken in the response to get the next page of results.

If you don't specify a value for MaxResults, Route 53 returns up to 100 configurations.

newListQueryLoggingConfigs :: ListQueryLoggingConfigs

Create a value of ListQueryLoggingConfigs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostedZoneId:ListQueryLoggingConfigs', listQueryLoggingConfigs_hostedZoneId - (Optional) If you want to list the query logging configuration that is associated with a hosted zone, specify the ID in HostedZoneId.

If you don't specify a hosted zone ID, ListQueryLoggingConfigs returns all of the configurations that are associated with the current Amazon Web Services account.

$sel:nextToken:ListQueryLoggingConfigs', listQueryLoggingConfigs_nextToken - (Optional) If the current Amazon Web Services account has more than MaxResults query logging configurations, use NextToken to get the second and subsequent pages of results.

For the first ListQueryLoggingConfigs request, omit this value.

For the second and subsequent requests, get the value of NextToken from the previous response and specify that value for NextToken in the request.

$sel:maxResults:ListQueryLoggingConfigs', listQueryLoggingConfigs_maxResults - (Optional) The maximum number of query logging configurations that you want Amazon Route 53 to return in response to the current request. If the current Amazon Web Services account has more than MaxResults configurations, use the value of NextToken in the response to get the next page of results.

If you don't specify a value for MaxResults, Route 53 returns up to 100 configurations.

Request Lenses

listQueryLoggingConfigs_hostedZoneId :: Lens' ListQueryLoggingConfigs (Maybe ResourceId)

(Optional) If you want to list the query logging configuration that is associated with a hosted zone, specify the ID in HostedZoneId.

If you don't specify a hosted zone ID, ListQueryLoggingConfigs returns all of the configurations that are associated with the current Amazon Web Services account.

listQueryLoggingConfigs_nextToken :: Lens' ListQueryLoggingConfigs (Maybe Text)

(Optional) If the current Amazon Web Services account has more than MaxResults query logging configurations, use NextToken to get the second and subsequent pages of results.

For the first ListQueryLoggingConfigs request, omit this value.

For the second and subsequent requests, get the value of NextToken from the previous response and specify that value for NextToken in the request.

listQueryLoggingConfigs_maxResults :: Lens' ListQueryLoggingConfigs (Maybe Text)

(Optional) The maximum number of query logging configurations that you want Amazon Route 53 to return in response to the current request. If the current Amazon Web Services account has more than MaxResults configurations, use the value of NextToken in the response to get the next page of results.

If you don't specify a value for MaxResults, Route 53 returns up to 100 configurations.

Destructuring the Response

data ListQueryLoggingConfigsResponse

See: newListQueryLoggingConfigsResponse smart constructor.

Constructors

ListQueryLoggingConfigsResponse'

Fields

nextToken :: Maybe Text

If a response includes the last of the query logging configurations that are associated with the current Amazon Web Services account, NextToken doesn't appear in the response.

If a response doesn't include the last of the configurations, you can get more configurations by submitting another ListQueryLoggingConfigs request. Get the value of NextToken that Amazon Route 53 returned in the previous response and include it in NextToken in the next request.

httpStatus :: Int

The response's http status code.

queryLoggingConfigs :: [QueryLoggingConfig]

An array that contains one QueryLoggingConfig element for each configuration for DNS query logging that is associated with the current Amazon Web Services account.

newListQueryLoggingConfigsResponse

Arguments

:: Int
$sel:httpStatus:ListQueryLoggingConfigsResponse'

-> ListQueryLoggingConfigsResponse

Create a value of ListQueryLoggingConfigsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListQueryLoggingConfigs', listQueryLoggingConfigsResponse_nextToken - If a response includes the last of the query logging configurations that are associated with the current Amazon Web Services account, NextToken doesn't appear in the response.

If a response doesn't include the last of the configurations, you can get more configurations by submitting another ListQueryLoggingConfigs request. Get the value of NextToken that Amazon Route 53 returned in the previous response and include it in NextToken in the next request.

$sel:httpStatus:ListQueryLoggingConfigsResponse', listQueryLoggingConfigsResponse_httpStatus - The response's http status code.

$sel:queryLoggingConfigs:ListQueryLoggingConfigsResponse', listQueryLoggingConfigsResponse_queryLoggingConfigs - An array that contains one QueryLoggingConfig element for each configuration for DNS query logging that is associated with the current Amazon Web Services account.

Response Lenses

listQueryLoggingConfigsResponse_nextToken :: Lens' ListQueryLoggingConfigsResponse (Maybe Text)

If a response includes the last of the query logging configurations that are associated with the current Amazon Web Services account, NextToken doesn't appear in the response.

If a response doesn't include the last of the configurations, you can get more configurations by submitting another ListQueryLoggingConfigs request. Get the value of NextToken that Amazon Route 53 returned in the previous response and include it in NextToken in the next request.

listQueryLoggingConfigsResponse_httpStatus :: Lens' ListQueryLoggingConfigsResponse Int

The response's http status code.

listQueryLoggingConfigsResponse_queryLoggingConfigs :: Lens' ListQueryLoggingConfigsResponse [QueryLoggingConfig]

An array that contains one QueryLoggingConfig element for each configuration for DNS query logging that is associated with the current Amazon Web Services account.