Network.AWS.Route53.ListResourceRecordSets

Description

Lists the resource record sets in a specified hosted zone.

ListResourceRecordSets returns up to 300 resource record sets at a time in ASCII order, beginning at a position specified by the name and type elements.

Sort order

ListResourceRecordSets sorts results first by DNS name with the labels reversed, for example:

com.example.www.

Note the trailing dot, which can change the sort order when the record name contains characters that appear before . (decimal 46) in the ASCII table. These characters include the following: ! " # $ % & ' ( ) * + , -

When multiple records have the same DNS name, ListResourceRecordSets sorts results by the record type.

Specifying where to start listing records

You can use the name and type elements to specify the resource record set that the list begins with:

If you do not specify Name or Type

The results begin with the first resource record set that the hosted zone contains.

If you specify Name but not Type

The results begin with the first resource record set in the list whose name is greater than or equal to Name.

If you specify Type but not Name

Amazon Route 53 returns the InvalidInput error.

If you specify both Name and Type

The results begin with the first resource record set in the list whose name is greater than or equal to Name, and whose type is greater than or equal to Type.

Resource record sets that are PENDING

This action returns the most current version of the records. This includes records that are PENDING, and that are not yet available on all Route 53 DNS servers.

Changing resource record sets

To ensure that you get an accurate listing of the resource record sets for a hosted zone at a point in time, do not submit a ChangeResourceRecordSets request while you're paging through the results of a ListResourceRecordSets request. If you do, some pages may display results without the latest changes while other pages display results with the latest changes.

Displaying the next page of results

If a ListResourceRecordSets command returns more than one page of results, the value of IsTruncated is true. To display the next page of results, get the values of NextRecordName, NextRecordType, and NextRecordIdentifier (if any) from the response. Then submit another ListResourceRecordSets request, and specify those values for StartRecordName, StartRecordType, and StartRecordIdentifier.

This operation returns paginated results.

Creating a Request

data ListResourceRecordSets

A request for the resource record sets that are associated with a specified hosted zone.

See: newListResourceRecordSets smart constructor.

Constructors

ListResourceRecordSets'

Fields

startRecordName :: Maybe Text

The first name in the lexicographic ordering of resource record sets that you want to list. If the specified record name doesn't exist, the results begin with the first resource record set that has a name greater than the value of name.

startRecordType :: Maybe RRType

The type of resource record set to begin the record listing from.

Valid values for basic resource record sets: A | AAAA | CAA | CNAME | MX | NAPTR | NS | PTR | SOA | SPF | SRV | TXT

Values for weighted, latency, geolocation, and failover resource record sets: A | AAAA | CAA | CNAME | MX | NAPTR | PTR | SPF | SRV | TXT

Values for alias resource record sets:

API Gateway custom regional API or edge-optimized API: A

CloudFront distribution: A or AAAA

Elastic Beanstalk environment that has a regionalized subdomain: A

Elastic Load Balancing load balancer: A | AAAA

S3 bucket: A

VPC interface VPC endpoint: A

Another resource record set in this hosted zone: The type of the resource record set that the alias references.

Constraint: Specifying type without specifying name returns an InvalidInput error.

startRecordIdentifier :: Maybe Text

Resource record sets that have a routing policy other than simple: If results were truncated for a given DNS name and type, specify the value of NextRecordIdentifier from the previous response to get the next resource record set that has the current DNS name and type.

maxItems :: Maybe Text

(Optional) The maximum number of resource records sets to include in the response body for this request. If the response includes more than maxitems resource record sets, the value of the IsTruncated element in the response is true, and the values of the NextRecordName and NextRecordType elements in the response identify the first resource record set in the next group of maxitems resource record sets.

hostedZoneId :: ResourceId

The ID of the hosted zone that contains the resource record sets that you want to list.

newListResourceRecordSets

Arguments

:: ResourceId
$sel:hostedZoneId:ListResourceRecordSets'

-> ListResourceRecordSets

Create a value of ListResourceRecordSets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startRecordName:ListResourceRecordSets', listResourceRecordSets_startRecordName - The first name in the lexicographic ordering of resource record sets that you want to list. If the specified record name doesn't exist, the results begin with the first resource record set that has a name greater than the value of name.

$sel:startRecordType:ListResourceRecordSets', listResourceRecordSets_startRecordType - The type of resource record set to begin the record listing from.

Valid values for basic resource record sets: A | AAAA | CAA | CNAME | MX | NAPTR | NS | PTR | SOA | SPF | SRV | TXT

Values for weighted, latency, geolocation, and failover resource record sets: A | AAAA | CAA | CNAME | MX | NAPTR | PTR | SPF | SRV | TXT

Values for alias resource record sets:

API Gateway custom regional API or edge-optimized API: A

CloudFront distribution: A or AAAA

Elastic Beanstalk environment that has a regionalized subdomain: A

Elastic Load Balancing load balancer: A | AAAA

S3 bucket: A

VPC interface VPC endpoint: A

Another resource record set in this hosted zone: The type of the resource record set that the alias references.

Constraint: Specifying type without specifying name returns an InvalidInput error.

$sel:startRecordIdentifier:ListResourceRecordSets', listResourceRecordSets_startRecordIdentifier - Resource record sets that have a routing policy other than simple: If results were truncated for a given DNS name and type, specify the value of NextRecordIdentifier from the previous response to get the next resource record set that has the current DNS name and type.

$sel:maxItems:ListResourceRecordSets', listResourceRecordSets_maxItems - (Optional) The maximum number of resource records sets to include in the response body for this request. If the response includes more than maxitems resource record sets, the value of the IsTruncated element in the response is true, and the values of the NextRecordName and NextRecordType elements in the response identify the first resource record set in the next group of maxitems resource record sets.

$sel:hostedZoneId:ListResourceRecordSets', listResourceRecordSets_hostedZoneId - The ID of the hosted zone that contains the resource record sets that you want to list.

Request Lenses

listResourceRecordSets_startRecordName :: Lens' ListResourceRecordSets (Maybe Text)

The first name in the lexicographic ordering of resource record sets that you want to list. If the specified record name doesn't exist, the results begin with the first resource record set that has a name greater than the value of name.

listResourceRecordSets_startRecordType :: Lens' ListResourceRecordSets (Maybe RRType)

The type of resource record set to begin the record listing from.

Valid values for basic resource record sets: A | AAAA | CAA | CNAME | MX | NAPTR | NS | PTR | SOA | SPF | SRV | TXT

Values for weighted, latency, geolocation, and failover resource record sets: A | AAAA | CAA | CNAME | MX | NAPTR | PTR | SPF | SRV | TXT

Values for alias resource record sets:

API Gateway custom regional API or edge-optimized API: A

CloudFront distribution: A or AAAA

Elastic Beanstalk environment that has a regionalized subdomain: A

Elastic Load Balancing load balancer: A | AAAA

S3 bucket: A

VPC interface VPC endpoint: A

Another resource record set in this hosted zone: The type of the resource record set that the alias references.

Constraint: Specifying type without specifying name returns an InvalidInput error.

listResourceRecordSets_startRecordIdentifier :: Lens' ListResourceRecordSets (Maybe Text)

Resource record sets that have a routing policy other than simple: If results were truncated for a given DNS name and type, specify the value of NextRecordIdentifier from the previous response to get the next resource record set that has the current DNS name and type.

listResourceRecordSets_maxItems :: Lens' ListResourceRecordSets (Maybe Text)

(Optional) The maximum number of resource records sets to include in the response body for this request. If the response includes more than maxitems resource record sets, the value of the IsTruncated element in the response is true, and the values of the NextRecordName and NextRecordType elements in the response identify the first resource record set in the next group of maxitems resource record sets.

listResourceRecordSets_hostedZoneId :: Lens' ListResourceRecordSets ResourceId

The ID of the hosted zone that contains the resource record sets that you want to list.

Destructuring the Response

data ListResourceRecordSetsResponse

A complex type that contains list information for the resource record set.

See: newListResourceRecordSetsResponse smart constructor.

Constructors

ListResourceRecordSetsResponse'

Fields

nextRecordType :: Maybe RRType

If the results were truncated, the type of the next record in the list.

This element is present only if IsTruncated is true.

nextRecordName :: Maybe Text

If the results were truncated, the name of the next record in the list.

This element is present only if IsTruncated is true.

nextRecordIdentifier :: Maybe Text

Resource record sets that have a routing policy other than simple: If results were truncated for a given DNS name and type, the value of SetIdentifier for the next resource record set that has the current DNS name and type.

For information about routing policies, see Choosing a Routing Policy in the Amazon Route 53 Developer Guide.

httpStatus :: Int

The response's http status code.

resourceRecordSets :: [ResourceRecordSet]

Information about multiple resource record sets.

isTruncated :: Bool

A flag that indicates whether more resource record sets remain to be listed. If your results were truncated, you can make a follow-up pagination request by using the NextRecordName element.

maxItems :: Text

The maximum number of records you requested.

newListResourceRecordSetsResponse

Arguments

:: Int
$sel:httpStatus:ListResourceRecordSetsResponse'

-> Bool
$sel:isTruncated:ListResourceRecordSetsResponse'

-> Text
$sel:maxItems:ListResourceRecordSets'

-> ListResourceRecordSetsResponse

Create a value of ListResourceRecordSetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextRecordType:ListResourceRecordSetsResponse', listResourceRecordSetsResponse_nextRecordType - If the results were truncated, the type of the next record in the list.

This element is present only if IsTruncated is true.

$sel:nextRecordName:ListResourceRecordSetsResponse', listResourceRecordSetsResponse_nextRecordName - If the results were truncated, the name of the next record in the list.

This element is present only if IsTruncated is true.

$sel:nextRecordIdentifier:ListResourceRecordSetsResponse', listResourceRecordSetsResponse_nextRecordIdentifier - Resource record sets that have a routing policy other than simple: If results were truncated for a given DNS name and type, the value of SetIdentifier for the next resource record set that has the current DNS name and type.

For information about routing policies, see Choosing a Routing Policy in the Amazon Route 53 Developer Guide.

$sel:httpStatus:ListResourceRecordSetsResponse', listResourceRecordSetsResponse_httpStatus - The response's http status code.

$sel:resourceRecordSets:ListResourceRecordSetsResponse', listResourceRecordSetsResponse_resourceRecordSets - Information about multiple resource record sets.

$sel:isTruncated:ListResourceRecordSetsResponse', listResourceRecordSetsResponse_isTruncated - A flag that indicates whether more resource record sets remain to be listed. If your results were truncated, you can make a follow-up pagination request by using the NextRecordName element.

$sel:maxItems:ListResourceRecordSets', listResourceRecordSetsResponse_maxItems - The maximum number of records you requested.

Response Lenses

listResourceRecordSetsResponse_nextRecordType :: Lens' ListResourceRecordSetsResponse (Maybe RRType)

If the results were truncated, the type of the next record in the list.

This element is present only if IsTruncated is true.

listResourceRecordSetsResponse_nextRecordName :: Lens' ListResourceRecordSetsResponse (Maybe Text)

If the results were truncated, the name of the next record in the list.

This element is present only if IsTruncated is true.

listResourceRecordSetsResponse_nextRecordIdentifier :: Lens' ListResourceRecordSetsResponse (Maybe Text)

Resource record sets that have a routing policy other than simple: If results were truncated for a given DNS name and type, the value of SetIdentifier for the next resource record set that has the current DNS name and type.

For information about routing policies, see Choosing a Routing Policy in the Amazon Route 53 Developer Guide.

listResourceRecordSetsResponse_httpStatus :: Lens' ListResourceRecordSetsResponse Int

The response's http status code.

listResourceRecordSetsResponse_resourceRecordSets :: Lens' ListResourceRecordSetsResponse [ResourceRecordSet]

Information about multiple resource record sets.

listResourceRecordSetsResponse_isTruncated :: Lens' ListResourceRecordSetsResponse Bool

A flag that indicates whether more resource record sets remain to be listed. If your results were truncated, you can make a follow-up pagination request by using the NextRecordName element.

listResourceRecordSetsResponse_maxItems :: Lens' ListResourceRecordSetsResponse Text

The maximum number of records you requested.