Network.AWS.Route53.ListTagsForResources

Description

Lists tags for up to 10 health checks or hosted zones.

For information about using tags for cost allocation, see Using Cost Allocation Tags in the Billing and Cost Management User Guide.

Creating a Request

data ListTagsForResources

A complex type that contains information about the health checks or hosted zones for which you want to list tags.

See: newListTagsForResources smart constructor.

Constructors

ListTagsForResources'

Fields

resourceType :: TagResourceType

The type of the resources.

The resource type for health checks is healthcheck.

The resource type for hosted zones is hostedzone.

resourceIds :: NonEmpty Text

A complex type that contains the ResourceId element for each resource for which you want to get a list of tags.

newListTagsForResources

Arguments

:: TagResourceType
$sel:resourceType:ListTagsForResources'

-> NonEmpty Text
$sel:resourceIds:ListTagsForResources'

-> ListTagsForResources

Create a value of ListTagsForResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ListTagsForResources', listTagsForResources_resourceType - The type of the resources.

The resource type for health checks is healthcheck.

The resource type for hosted zones is hostedzone.

$sel:resourceIds:ListTagsForResources', listTagsForResources_resourceIds - A complex type that contains the ResourceId element for each resource for which you want to get a list of tags.

Request Lenses

listTagsForResources_resourceType :: Lens' ListTagsForResources TagResourceType

The type of the resources.

The resource type for health checks is healthcheck.

The resource type for hosted zones is hostedzone.

listTagsForResources_resourceIds :: Lens' ListTagsForResources (NonEmpty Text)

A complex type that contains the ResourceId element for each resource for which you want to get a list of tags.

Destructuring the Response

data ListTagsForResourcesResponse

A complex type containing tags for the specified resources.

See: newListTagsForResourcesResponse smart constructor.

Constructors

ListTagsForResourcesResponse'

Fields

httpStatus :: Int

The response's http status code.

resourceTagSets :: [ResourceTagSet]

A list of ResourceTagSets containing tags associated with the specified resources.

newListTagsForResourcesResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourcesResponse'

-> ListTagsForResourcesResponse

Create a value of ListTagsForResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ListTagsForResourcesResponse', listTagsForResourcesResponse_httpStatus - The response's http status code.

$sel:resourceTagSets:ListTagsForResourcesResponse', listTagsForResourcesResponse_resourceTagSets - A list of ResourceTagSets containing tags associated with the specified resources.

Response Lenses

listTagsForResourcesResponse_httpStatus :: Lens' ListTagsForResourcesResponse Int

The response's http status code.

listTagsForResourcesResponse_resourceTagSets :: Lens' ListTagsForResourcesResponse [ResourceTagSet]

A list of ResourceTagSets containing tags associated with the specified resources.