Network.AWS.Route53.ListTrafficPolicies

Description

Gets information about the latest version for every traffic policy that is associated with the current Amazon Web Services account. Policies are listed in the order that they were created in.

For information about how of deleting a traffic policy affects the response from ListTrafficPolicies, see DeleteTrafficPolicy.

Creating a Request

data ListTrafficPolicies

A complex type that contains the information about the request to list the traffic policies that are associated with the current Amazon Web Services account.

See: newListTrafficPolicies smart constructor.

Constructors

ListTrafficPolicies'

Fields

trafficPolicyIdMarker :: Maybe Text

(Conditional) For your first request to ListTrafficPolicies, don't include the TrafficPolicyIdMarker parameter.

If you have more traffic policies than the value of MaxItems, ListTrafficPolicies returns only the first MaxItems traffic policies. To get the next group of policies, submit another request to ListTrafficPolicies. For the value of TrafficPolicyIdMarker, specify the value of TrafficPolicyIdMarker that was returned in the previous response.

maxItems :: Maybe Text

(Optional) The maximum number of traffic policies that you want Amazon Route 53 to return in response to this request. If you have more than MaxItems traffic policies, the value of IsTruncated in the response is true, and the value of TrafficPolicyIdMarker is the ID of the first traffic policy that Route 53 will return if you submit another request.

newListTrafficPolicies :: ListTrafficPolicies

Create a value of ListTrafficPolicies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trafficPolicyIdMarker:ListTrafficPolicies', listTrafficPolicies_trafficPolicyIdMarker - (Conditional) For your first request to ListTrafficPolicies, don't include the TrafficPolicyIdMarker parameter.

If you have more traffic policies than the value of MaxItems, ListTrafficPolicies returns only the first MaxItems traffic policies. To get the next group of policies, submit another request to ListTrafficPolicies. For the value of TrafficPolicyIdMarker, specify the value of TrafficPolicyIdMarker that was returned in the previous response.

$sel:maxItems:ListTrafficPolicies', listTrafficPolicies_maxItems - (Optional) The maximum number of traffic policies that you want Amazon Route 53 to return in response to this request. If you have more than MaxItems traffic policies, the value of IsTruncated in the response is true, and the value of TrafficPolicyIdMarker is the ID of the first traffic policy that Route 53 will return if you submit another request.

Request Lenses

listTrafficPolicies_trafficPolicyIdMarker :: Lens' ListTrafficPolicies (Maybe Text)

(Conditional) For your first request to ListTrafficPolicies, don't include the TrafficPolicyIdMarker parameter.

If you have more traffic policies than the value of MaxItems, ListTrafficPolicies returns only the first MaxItems traffic policies. To get the next group of policies, submit another request to ListTrafficPolicies. For the value of TrafficPolicyIdMarker, specify the value of TrafficPolicyIdMarker that was returned in the previous response.

listTrafficPolicies_maxItems :: Lens' ListTrafficPolicies (Maybe Text)

(Optional) The maximum number of traffic policies that you want Amazon Route 53 to return in response to this request. If you have more than MaxItems traffic policies, the value of IsTruncated in the response is true, and the value of TrafficPolicyIdMarker is the ID of the first traffic policy that Route 53 will return if you submit another request.

Destructuring the Response

data ListTrafficPoliciesResponse

A complex type that contains the response information for the request.

See: newListTrafficPoliciesResponse smart constructor.

Constructors

ListTrafficPoliciesResponse'

Fields

httpStatus :: Int

The response's http status code.

trafficPolicySummaries :: [TrafficPolicySummary]

A list that contains one TrafficPolicySummary element for each traffic policy that was created by the current Amazon Web Services account.

isTruncated :: Bool

A flag that indicates whether there are more traffic policies to be listed. If the response was truncated, you can get the next group of traffic policies by submitting another ListTrafficPolicies request and specifying the value of TrafficPolicyIdMarker in the TrafficPolicyIdMarker request parameter.

trafficPolicyIdMarker :: Text

If the value of IsTruncated is true, TrafficPolicyIdMarker is the ID of the first traffic policy in the next group of MaxItems traffic policies.

maxItems :: Text

The value that you specified for the MaxItems parameter in the ListTrafficPolicies request that produced the current response.

newListTrafficPoliciesResponse

Arguments

:: Int
$sel:httpStatus:ListTrafficPoliciesResponse'

-> Bool
$sel:isTruncated:ListTrafficPoliciesResponse'

-> Text
$sel:trafficPolicyIdMarker:ListTrafficPolicies'

-> Text
$sel:maxItems:ListTrafficPolicies'

-> ListTrafficPoliciesResponse

Create a value of ListTrafficPoliciesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ListTrafficPoliciesResponse', listTrafficPoliciesResponse_httpStatus - The response's http status code.

$sel:trafficPolicySummaries:ListTrafficPoliciesResponse', listTrafficPoliciesResponse_trafficPolicySummaries - A list that contains one TrafficPolicySummary element for each traffic policy that was created by the current Amazon Web Services account.

$sel:isTruncated:ListTrafficPoliciesResponse', listTrafficPoliciesResponse_isTruncated - A flag that indicates whether there are more traffic policies to be listed. If the response was truncated, you can get the next group of traffic policies by submitting another ListTrafficPolicies request and specifying the value of TrafficPolicyIdMarker in the TrafficPolicyIdMarker request parameter.

$sel:trafficPolicyIdMarker:ListTrafficPolicies', listTrafficPoliciesResponse_trafficPolicyIdMarker - If the value of IsTruncated is true, TrafficPolicyIdMarker is the ID of the first traffic policy in the next group of MaxItems traffic policies.

$sel:maxItems:ListTrafficPolicies', listTrafficPoliciesResponse_maxItems - The value that you specified for the MaxItems parameter in the ListTrafficPolicies request that produced the current response.

Response Lenses

listTrafficPoliciesResponse_httpStatus :: Lens' ListTrafficPoliciesResponse Int

The response's http status code.

listTrafficPoliciesResponse_trafficPolicySummaries :: Lens' ListTrafficPoliciesResponse [TrafficPolicySummary]

A list that contains one TrafficPolicySummary element for each traffic policy that was created by the current Amazon Web Services account.

listTrafficPoliciesResponse_isTruncated :: Lens' ListTrafficPoliciesResponse Bool

A flag that indicates whether there are more traffic policies to be listed. If the response was truncated, you can get the next group of traffic policies by submitting another ListTrafficPolicies request and specifying the value of TrafficPolicyIdMarker in the TrafficPolicyIdMarker request parameter.

listTrafficPoliciesResponse_trafficPolicyIdMarker :: Lens' ListTrafficPoliciesResponse Text

If the value of IsTruncated is true, TrafficPolicyIdMarker is the ID of the first traffic policy in the next group of MaxItems traffic policies.

listTrafficPoliciesResponse_maxItems :: Lens' ListTrafficPoliciesResponse Text

The value that you specified for the MaxItems parameter in the ListTrafficPolicies request that produced the current response.