Network.AWS.Route53.ListTrafficPolicyInstancesByHostedZone

Description

Gets information about the traffic policy instances that you created in a specified hosted zone.

After you submit a CreateTrafficPolicyInstance or an UpdateTrafficPolicyInstance request, there's a brief delay while Amazon Route 53 creates the resource record sets that are specified in the traffic policy definition. For more information, see the State response element.

Route 53 returns a maximum of 100 items in each response. If you have a lot of traffic policy instances, you can use the MaxItems parameter to list them in groups of up to 100.

Creating a Request

data ListTrafficPolicyInstancesByHostedZone

A request for the traffic policy instances that you created in a specified hosted zone.

See: newListTrafficPolicyInstancesByHostedZone smart constructor.

Constructors

ListTrafficPolicyInstancesByHostedZone'

Fields

trafficPolicyInstanceTypeMarker :: Maybe RRType

If the value of IsTruncated in the previous response is true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancetype, specify the value of TrafficPolicyInstanceTypeMarker from the previous response, which is the type of the first traffic policy instance in the next group of traffic policy instances.

If the value of IsTruncated in the previous response was false, there are no more traffic policy instances to get.

maxItems :: Maybe Text

The maximum number of traffic policy instances to be included in the response body for this request. If you have more than MaxItems traffic policy instances, the value of the IsTruncated element in the response is true, and the values of HostedZoneIdMarker, TrafficPolicyInstanceNameMarker, and TrafficPolicyInstanceTypeMarker represent the first traffic policy instance that Amazon Route 53 will return if you submit another request.

trafficPolicyInstanceNameMarker :: Maybe Text

If the value of IsTruncated in the previous response is true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancename, specify the value of TrafficPolicyInstanceNameMarker from the previous response, which is the name of the first traffic policy instance in the next group of traffic policy instances.

If the value of IsTruncated in the previous response was false, there are no more traffic policy instances to get.

hostedZoneId :: ResourceId

The ID of the hosted zone that you want to list traffic policy instances for.

newListTrafficPolicyInstancesByHostedZone

Arguments

:: ResourceId
$sel:hostedZoneId:ListTrafficPolicyInstancesByHostedZone'

-> ListTrafficPolicyInstancesByHostedZone

Create a value of ListTrafficPolicyInstancesByHostedZone with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trafficPolicyInstanceTypeMarker:ListTrafficPolicyInstancesByHostedZone', listTrafficPolicyInstancesByHostedZone_trafficPolicyInstanceTypeMarker - If the value of IsTruncated in the previous response is true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancetype, specify the value of TrafficPolicyInstanceTypeMarker from the previous response, which is the type of the first traffic policy instance in the next group of traffic policy instances.

If the value of IsTruncated in the previous response was false, there are no more traffic policy instances to get.

$sel:maxItems:ListTrafficPolicyInstancesByHostedZone', listTrafficPolicyInstancesByHostedZone_maxItems - The maximum number of traffic policy instances to be included in the response body for this request. If you have more than MaxItems traffic policy instances, the value of the IsTruncated element in the response is true, and the values of HostedZoneIdMarker, TrafficPolicyInstanceNameMarker, and TrafficPolicyInstanceTypeMarker represent the first traffic policy instance that Amazon Route 53 will return if you submit another request.

$sel:trafficPolicyInstanceNameMarker:ListTrafficPolicyInstancesByHostedZone', listTrafficPolicyInstancesByHostedZone_trafficPolicyInstanceNameMarker - If the value of IsTruncated in the previous response is true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancename, specify the value of TrafficPolicyInstanceNameMarker from the previous response, which is the name of the first traffic policy instance in the next group of traffic policy instances.

If the value of IsTruncated in the previous response was false, there are no more traffic policy instances to get.

$sel:hostedZoneId:ListTrafficPolicyInstancesByHostedZone', listTrafficPolicyInstancesByHostedZone_hostedZoneId - The ID of the hosted zone that you want to list traffic policy instances for.

Request Lenses

listTrafficPolicyInstancesByHostedZone_trafficPolicyInstanceTypeMarker :: Lens' ListTrafficPolicyInstancesByHostedZone (Maybe RRType)

If the value of IsTruncated in the previous response is true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancetype, specify the value of TrafficPolicyInstanceTypeMarker from the previous response, which is the type of the first traffic policy instance in the next group of traffic policy instances.

If the value of IsTruncated in the previous response was false, there are no more traffic policy instances to get.

listTrafficPolicyInstancesByHostedZone_maxItems :: Lens' ListTrafficPolicyInstancesByHostedZone (Maybe Text)

The maximum number of traffic policy instances to be included in the response body for this request. If you have more than MaxItems traffic policy instances, the value of the IsTruncated element in the response is true, and the values of HostedZoneIdMarker, TrafficPolicyInstanceNameMarker, and TrafficPolicyInstanceTypeMarker represent the first traffic policy instance that Amazon Route 53 will return if you submit another request.

listTrafficPolicyInstancesByHostedZone_trafficPolicyInstanceNameMarker :: Lens' ListTrafficPolicyInstancesByHostedZone (Maybe Text)

If the value of IsTruncated in the previous response is true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancename, specify the value of TrafficPolicyInstanceNameMarker from the previous response, which is the name of the first traffic policy instance in the next group of traffic policy instances.

If the value of IsTruncated in the previous response was false, there are no more traffic policy instances to get.

listTrafficPolicyInstancesByHostedZone_hostedZoneId :: Lens' ListTrafficPolicyInstancesByHostedZone ResourceId

The ID of the hosted zone that you want to list traffic policy instances for.

Destructuring the Response

data ListTrafficPolicyInstancesByHostedZoneResponse

A complex type that contains the response information for the request.

See: newListTrafficPolicyInstancesByHostedZoneResponse smart constructor.

Constructors

ListTrafficPolicyInstancesByHostedZoneResponse'

Fields

trafficPolicyInstanceTypeMarker :: Maybe RRType

If IsTruncated is true, TrafficPolicyInstanceTypeMarker is the DNS type of the resource record sets that are associated with the first traffic policy instance in the next group of traffic policy instances.

trafficPolicyInstanceNameMarker :: Maybe Text

If IsTruncated is true, TrafficPolicyInstanceNameMarker is the name of the first traffic policy instance in the next group of traffic policy instances.

httpStatus :: Int

The response's http status code.

trafficPolicyInstances :: [TrafficPolicyInstance]

A list that contains one TrafficPolicyInstance element for each traffic policy instance that matches the elements in the request.

isTruncated :: Bool

A flag that indicates whether there are more traffic policy instances to be listed. If the response was truncated, you can get the next group of traffic policy instances by submitting another ListTrafficPolicyInstancesByHostedZone request and specifying the values of HostedZoneIdMarker, TrafficPolicyInstanceNameMarker, and TrafficPolicyInstanceTypeMarker in the corresponding request parameters.

maxItems :: Text

The value that you specified for the MaxItems parameter in the ListTrafficPolicyInstancesByHostedZone request that produced the current response.

newListTrafficPolicyInstancesByHostedZoneResponse

Arguments

:: Int
$sel:httpStatus:ListTrafficPolicyInstancesByHostedZoneResponse'

-> Bool
$sel:isTruncated:ListTrafficPolicyInstancesByHostedZoneResponse'

-> Text
$sel:maxItems:ListTrafficPolicyInstancesByHostedZone'

-> ListTrafficPolicyInstancesByHostedZoneResponse

Create a value of ListTrafficPolicyInstancesByHostedZoneResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trafficPolicyInstanceTypeMarker:ListTrafficPolicyInstancesByHostedZone', listTrafficPolicyInstancesByHostedZoneResponse_trafficPolicyInstanceTypeMarker - If IsTruncated is true, TrafficPolicyInstanceTypeMarker is the DNS type of the resource record sets that are associated with the first traffic policy instance in the next group of traffic policy instances.

$sel:trafficPolicyInstanceNameMarker:ListTrafficPolicyInstancesByHostedZone', listTrafficPolicyInstancesByHostedZoneResponse_trafficPolicyInstanceNameMarker - If IsTruncated is true, TrafficPolicyInstanceNameMarker is the name of the first traffic policy instance in the next group of traffic policy instances.

$sel:httpStatus:ListTrafficPolicyInstancesByHostedZoneResponse', listTrafficPolicyInstancesByHostedZoneResponse_httpStatus - The response's http status code.

$sel:trafficPolicyInstances:ListTrafficPolicyInstancesByHostedZoneResponse', listTrafficPolicyInstancesByHostedZoneResponse_trafficPolicyInstances - A list that contains one TrafficPolicyInstance element for each traffic policy instance that matches the elements in the request.

$sel:isTruncated:ListTrafficPolicyInstancesByHostedZoneResponse', listTrafficPolicyInstancesByHostedZoneResponse_isTruncated - A flag that indicates whether there are more traffic policy instances to be listed. If the response was truncated, you can get the next group of traffic policy instances by submitting another ListTrafficPolicyInstancesByHostedZone request and specifying the values of HostedZoneIdMarker, TrafficPolicyInstanceNameMarker, and TrafficPolicyInstanceTypeMarker in the corresponding request parameters.

$sel:maxItems:ListTrafficPolicyInstancesByHostedZone', listTrafficPolicyInstancesByHostedZoneResponse_maxItems - The value that you specified for the MaxItems parameter in the ListTrafficPolicyInstancesByHostedZone request that produced the current response.

Response Lenses

listTrafficPolicyInstancesByHostedZoneResponse_trafficPolicyInstanceTypeMarker :: Lens' ListTrafficPolicyInstancesByHostedZoneResponse (Maybe RRType)

If IsTruncated is true, TrafficPolicyInstanceTypeMarker is the DNS type of the resource record sets that are associated with the first traffic policy instance in the next group of traffic policy instances.

listTrafficPolicyInstancesByHostedZoneResponse_trafficPolicyInstanceNameMarker :: Lens' ListTrafficPolicyInstancesByHostedZoneResponse (Maybe Text)

If IsTruncated is true, TrafficPolicyInstanceNameMarker is the name of the first traffic policy instance in the next group of traffic policy instances.

listTrafficPolicyInstancesByHostedZoneResponse_httpStatus :: Lens' ListTrafficPolicyInstancesByHostedZoneResponse Int

The response's http status code.

listTrafficPolicyInstancesByHostedZoneResponse_trafficPolicyInstances :: Lens' ListTrafficPolicyInstancesByHostedZoneResponse [TrafficPolicyInstance]

A list that contains one TrafficPolicyInstance element for each traffic policy instance that matches the elements in the request.

listTrafficPolicyInstancesByHostedZoneResponse_isTruncated :: Lens' ListTrafficPolicyInstancesByHostedZoneResponse Bool

A flag that indicates whether there are more traffic policy instances to be listed. If the response was truncated, you can get the next group of traffic policy instances by submitting another ListTrafficPolicyInstancesByHostedZone request and specifying the values of HostedZoneIdMarker, TrafficPolicyInstanceNameMarker, and TrafficPolicyInstanceTypeMarker in the corresponding request parameters.

listTrafficPolicyInstancesByHostedZoneResponse_maxItems :: Lens' ListTrafficPolicyInstancesByHostedZoneResponse Text

The value that you specified for the MaxItems parameter in the ListTrafficPolicyInstancesByHostedZone request that produced the current response.