Network.AWS.Route53.ListTrafficPolicyVersions

Description

Gets information about all of the versions for a specified traffic policy.

Traffic policy versions are listed in numerical order by VersionNumber.

Creating a Request

data ListTrafficPolicyVersions

A complex type that contains the information about the request to list your traffic policies.

See: newListTrafficPolicyVersions smart constructor.

Constructors

ListTrafficPolicyVersions'

Fields

trafficPolicyVersionMarker :: Maybe Text

For your first request to ListTrafficPolicyVersions, don't include the TrafficPolicyVersionMarker parameter.

If you have more traffic policy versions than the value of MaxItems, ListTrafficPolicyVersions returns only the first group of MaxItems versions. To get more traffic policy versions, submit another ListTrafficPolicyVersions request. For the value of TrafficPolicyVersionMarker, specify the value of TrafficPolicyVersionMarker in the previous response.

maxItems :: Maybe Text

The maximum number of traffic policy versions that you want Amazon Route 53 to include in the response body for this request. If the specified traffic policy has more than MaxItems versions, the value of IsTruncated in the response is true, and the value of the TrafficPolicyVersionMarker element is the ID of the first version that Route 53 will return if you submit another request.

id :: Text

Specify the value of Id of the traffic policy for which you want to list all versions.

newListTrafficPolicyVersions

Arguments

:: Text
$sel:id:ListTrafficPolicyVersions'

-> ListTrafficPolicyVersions

Create a value of ListTrafficPolicyVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trafficPolicyVersionMarker:ListTrafficPolicyVersions', listTrafficPolicyVersions_trafficPolicyVersionMarker - For your first request to ListTrafficPolicyVersions, don't include the TrafficPolicyVersionMarker parameter.

If you have more traffic policy versions than the value of MaxItems, ListTrafficPolicyVersions returns only the first group of MaxItems versions. To get more traffic policy versions, submit another ListTrafficPolicyVersions request. For the value of TrafficPolicyVersionMarker, specify the value of TrafficPolicyVersionMarker in the previous response.

$sel:maxItems:ListTrafficPolicyVersions', listTrafficPolicyVersions_maxItems - The maximum number of traffic policy versions that you want Amazon Route 53 to include in the response body for this request. If the specified traffic policy has more than MaxItems versions, the value of IsTruncated in the response is true, and the value of the TrafficPolicyVersionMarker element is the ID of the first version that Route 53 will return if you submit another request.

$sel:id:ListTrafficPolicyVersions', listTrafficPolicyVersions_id - Specify the value of Id of the traffic policy for which you want to list all versions.

Request Lenses

listTrafficPolicyVersions_trafficPolicyVersionMarker :: Lens' ListTrafficPolicyVersions (Maybe Text)

For your first request to ListTrafficPolicyVersions, don't include the TrafficPolicyVersionMarker parameter.

If you have more traffic policy versions than the value of MaxItems, ListTrafficPolicyVersions returns only the first group of MaxItems versions. To get more traffic policy versions, submit another ListTrafficPolicyVersions request. For the value of TrafficPolicyVersionMarker, specify the value of TrafficPolicyVersionMarker in the previous response.

listTrafficPolicyVersions_maxItems :: Lens' ListTrafficPolicyVersions (Maybe Text)

The maximum number of traffic policy versions that you want Amazon Route 53 to include in the response body for this request. If the specified traffic policy has more than MaxItems versions, the value of IsTruncated in the response is true, and the value of the TrafficPolicyVersionMarker element is the ID of the first version that Route 53 will return if you submit another request.

listTrafficPolicyVersions_id :: Lens' ListTrafficPolicyVersions Text

Specify the value of Id of the traffic policy for which you want to list all versions.

Destructuring the Response

data ListTrafficPolicyVersionsResponse

A complex type that contains the response information for the request.

See: newListTrafficPolicyVersionsResponse smart constructor.

Constructors

ListTrafficPolicyVersionsResponse'

Fields

httpStatus :: Int

The response's http status code.

trafficPolicies :: [TrafficPolicy]

A list that contains one TrafficPolicy element for each traffic policy version that is associated with the specified traffic policy.

isTruncated :: Bool

A flag that indicates whether there are more traffic policies to be listed. If the response was truncated, you can get the next group of traffic policies by submitting another ListTrafficPolicyVersions request and specifying the value of NextMarker in the marker parameter.

trafficPolicyVersionMarker :: Text

If IsTruncated is true, the value of TrafficPolicyVersionMarker identifies the first traffic policy that Amazon Route 53 will return if you submit another request. Call ListTrafficPolicyVersions again and specify the value of TrafficPolicyVersionMarker in the TrafficPolicyVersionMarker request parameter.

This element is present only if IsTruncated is true.

maxItems :: Text

The value that you specified for the maxitems parameter in the ListTrafficPolicyVersions request that produced the current response.

newListTrafficPolicyVersionsResponse

Arguments

:: Int
$sel:httpStatus:ListTrafficPolicyVersionsResponse'

-> Bool
$sel:isTruncated:ListTrafficPolicyVersionsResponse'

-> Text
$sel:trafficPolicyVersionMarker:ListTrafficPolicyVersions'

-> Text
$sel:maxItems:ListTrafficPolicyVersions'

-> ListTrafficPolicyVersionsResponse

Create a value of ListTrafficPolicyVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ListTrafficPolicyVersionsResponse', listTrafficPolicyVersionsResponse_httpStatus - The response's http status code.

$sel:trafficPolicies:ListTrafficPolicyVersionsResponse', listTrafficPolicyVersionsResponse_trafficPolicies - A list that contains one TrafficPolicy element for each traffic policy version that is associated with the specified traffic policy.

$sel:isTruncated:ListTrafficPolicyVersionsResponse', listTrafficPolicyVersionsResponse_isTruncated - A flag that indicates whether there are more traffic policies to be listed. If the response was truncated, you can get the next group of traffic policies by submitting another ListTrafficPolicyVersions request and specifying the value of NextMarker in the marker parameter.

$sel:trafficPolicyVersionMarker:ListTrafficPolicyVersions', listTrafficPolicyVersionsResponse_trafficPolicyVersionMarker - If IsTruncated is true, the value of TrafficPolicyVersionMarker identifies the first traffic policy that Amazon Route 53 will return if you submit another request. Call ListTrafficPolicyVersions again and specify the value of TrafficPolicyVersionMarker in the TrafficPolicyVersionMarker request parameter.

This element is present only if IsTruncated is true.

$sel:maxItems:ListTrafficPolicyVersions', listTrafficPolicyVersionsResponse_maxItems - The value that you specified for the maxitems parameter in the ListTrafficPolicyVersions request that produced the current response.

Response Lenses

listTrafficPolicyVersionsResponse_httpStatus :: Lens' ListTrafficPolicyVersionsResponse Int

The response's http status code.

listTrafficPolicyVersionsResponse_trafficPolicies :: Lens' ListTrafficPolicyVersionsResponse [TrafficPolicy]

A list that contains one TrafficPolicy element for each traffic policy version that is associated with the specified traffic policy.

listTrafficPolicyVersionsResponse_isTruncated :: Lens' ListTrafficPolicyVersionsResponse Bool

A flag that indicates whether there are more traffic policies to be listed. If the response was truncated, you can get the next group of traffic policies by submitting another ListTrafficPolicyVersions request and specifying the value of NextMarker in the marker parameter.

listTrafficPolicyVersionsResponse_trafficPolicyVersionMarker :: Lens' ListTrafficPolicyVersionsResponse Text

If IsTruncated is true, the value of TrafficPolicyVersionMarker identifies the first traffic policy that Amazon Route 53 will return if you submit another request. Call ListTrafficPolicyVersions again and specify the value of TrafficPolicyVersionMarker in the TrafficPolicyVersionMarker request parameter.

This element is present only if IsTruncated is true.

listTrafficPolicyVersionsResponse_maxItems :: Lens' ListTrafficPolicyVersionsResponse Text

The value that you specified for the maxitems parameter in the ListTrafficPolicyVersions request that produced the current response.