Network.AWS.Route53.ListVPCAssociationAuthorizations

Description

Gets a list of the VPCs that were created by other accounts and that can be associated with a specified hosted zone because you've submitted one or more CreateVPCAssociationAuthorization requests.

The response includes a VPCs element with a VPC child element for each VPC that can be associated with the hosted zone.

This operation returns paginated results.

Creating a Request

data ListVPCAssociationAuthorizations

A complex type that contains information about that can be associated with your hosted zone.

See: newListVPCAssociationAuthorizations smart constructor.

Constructors

ListVPCAssociationAuthorizations'

Fields

nextToken :: Maybe Text

Optional: If a response includes a NextToken element, there are more VPCs that can be associated with the specified hosted zone. To get the next page of results, submit another request, and include the value of NextToken from the response in the nexttoken parameter in another ListVPCAssociationAuthorizations request.

maxResults :: Maybe Text

Optional: An integer that specifies the maximum number of VPCs that you want Amazon Route 53 to return. If you don't specify a value for MaxResults, Route 53 returns up to 50 VPCs per page.

hostedZoneId :: ResourceId

The ID of the hosted zone for which you want a list of VPCs that can be associated with the hosted zone.

newListVPCAssociationAuthorizations

Arguments

:: ResourceId
$sel:hostedZoneId:ListVPCAssociationAuthorizations'

-> ListVPCAssociationAuthorizations

Create a value of ListVPCAssociationAuthorizations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListVPCAssociationAuthorizations', listVPCAssociationAuthorizations_nextToken - Optional: If a response includes a NextToken element, there are more VPCs that can be associated with the specified hosted zone. To get the next page of results, submit another request, and include the value of NextToken from the response in the nexttoken parameter in another ListVPCAssociationAuthorizations request.

$sel:maxResults:ListVPCAssociationAuthorizations', listVPCAssociationAuthorizations_maxResults - Optional: An integer that specifies the maximum number of VPCs that you want Amazon Route 53 to return. If you don't specify a value for MaxResults, Route 53 returns up to 50 VPCs per page.

$sel:hostedZoneId:ListVPCAssociationAuthorizations', listVPCAssociationAuthorizations_hostedZoneId - The ID of the hosted zone for which you want a list of VPCs that can be associated with the hosted zone.

Request Lenses

listVPCAssociationAuthorizations_nextToken :: Lens' ListVPCAssociationAuthorizations (Maybe Text)

Optional: If a response includes a NextToken element, there are more VPCs that can be associated with the specified hosted zone. To get the next page of results, submit another request, and include the value of NextToken from the response in the nexttoken parameter in another ListVPCAssociationAuthorizations request.

listVPCAssociationAuthorizations_maxResults :: Lens' ListVPCAssociationAuthorizations (Maybe Text)

Optional: An integer that specifies the maximum number of VPCs that you want Amazon Route 53 to return. If you don't specify a value for MaxResults, Route 53 returns up to 50 VPCs per page.

listVPCAssociationAuthorizations_hostedZoneId :: Lens' ListVPCAssociationAuthorizations ResourceId

The ID of the hosted zone for which you want a list of VPCs that can be associated with the hosted zone.

Destructuring the Response

data ListVPCAssociationAuthorizationsResponse

A complex type that contains the response information for the request.

See: newListVPCAssociationAuthorizationsResponse smart constructor.

Constructors

ListVPCAssociationAuthorizationsResponse'

Fields

nextToken :: Maybe Text

When the response includes a NextToken element, there are more VPCs that can be associated with the specified hosted zone. To get the next page of VPCs, submit another ListVPCAssociationAuthorizations request, and include the value of the NextToken element from the response in the nexttoken request parameter.

httpStatus :: Int

The response's http status code.

hostedZoneId :: ResourceId

The ID of the hosted zone that you can associate the listed VPCs with.

vPCs :: NonEmpty VPC

The list of VPCs that are authorized to be associated with the specified hosted zone.

newListVPCAssociationAuthorizationsResponse

Arguments

:: Int
$sel:httpStatus:ListVPCAssociationAuthorizationsResponse'

-> ResourceId
$sel:hostedZoneId:ListVPCAssociationAuthorizations'

-> NonEmpty VPC
$sel:vPCs:ListVPCAssociationAuthorizationsResponse'

-> ListVPCAssociationAuthorizationsResponse

Create a value of ListVPCAssociationAuthorizationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListVPCAssociationAuthorizations', listVPCAssociationAuthorizationsResponse_nextToken - When the response includes a NextToken element, there are more VPCs that can be associated with the specified hosted zone. To get the next page of VPCs, submit another ListVPCAssociationAuthorizations request, and include the value of the NextToken element from the response in the nexttoken request parameter.

$sel:httpStatus:ListVPCAssociationAuthorizationsResponse', listVPCAssociationAuthorizationsResponse_httpStatus - The response's http status code.

$sel:hostedZoneId:ListVPCAssociationAuthorizations', listVPCAssociationAuthorizationsResponse_hostedZoneId - The ID of the hosted zone that you can associate the listed VPCs with.

$sel:vPCs:ListVPCAssociationAuthorizationsResponse', listVPCAssociationAuthorizationsResponse_vPCs - The list of VPCs that are authorized to be associated with the specified hosted zone.

Response Lenses

listVPCAssociationAuthorizationsResponse_nextToken :: Lens' ListVPCAssociationAuthorizationsResponse (Maybe Text)

When the response includes a NextToken element, there are more VPCs that can be associated with the specified hosted zone. To get the next page of VPCs, submit another ListVPCAssociationAuthorizations request, and include the value of the NextToken element from the response in the nexttoken request parameter.

listVPCAssociationAuthorizationsResponse_httpStatus :: Lens' ListVPCAssociationAuthorizationsResponse Int

The response's http status code.

listVPCAssociationAuthorizationsResponse_hostedZoneId :: Lens' ListVPCAssociationAuthorizationsResponse ResourceId

The ID of the hosted zone that you can associate the listed VPCs with.

listVPCAssociationAuthorizationsResponse_vPCs :: Lens' ListVPCAssociationAuthorizationsResponse (NonEmpty VPC)

The list of VPCs that are authorized to be associated with the specified hosted zone.