Network.AWS.Route53.TestDNSAnswer

Description

Gets the value that Amazon Route 53 returns in response to a DNS request for a specified record name and type. You can optionally specify the IP address of a DNS resolver, an EDNS0 client subnet IP address, and a subnet mask.

This call only supports querying public hosted zones.

Creating a Request

data TestDNSAnswer

Gets the value that Amazon Route 53 returns in response to a DNS request for a specified record name and type. You can optionally specify the IP address of a DNS resolver, an EDNS0 client subnet IP address, and a subnet mask.

See: newTestDNSAnswer smart constructor.

Constructors

TestDNSAnswer'

Fields

resolverIP :: Maybe Text

If you want to simulate a request from a specific DNS resolver, specify the IP address for that resolver. If you omit this value, TestDnsAnswer uses the IP address of a DNS resolver in the Amazon Web Services US East (N. Virginia) Region (us-east-1).

eDNS0ClientSubnetIP :: Maybe Text

If the resolver that you specified for resolverip supports EDNS0, specify the IPv4 or IPv6 address of a client in the applicable location, for example, 192.0.2.44 or 2001:db8:85a3::8a2e:370:7334.

eDNS0ClientSubnetMask :: Maybe Text

If you specify an IP address for edns0clientsubnetip, you can optionally specify the number of bits of the IP address that you want the checking tool to include in the DNS query. For example, if you specify 192.0.2.44 for edns0clientsubnetip and 24 for edns0clientsubnetmask, the checking tool will simulate a request from 192.0.2.0/24. The default value is 24 bits for IPv4 addresses and 64 bits for IPv6 addresses.

The range of valid values depends on whether edns0clientsubnetip is an IPv4 or an IPv6 address:

IPv4: Specify a value between 0 and 32

IPv6: Specify a value between 0 and 128

hostedZoneId :: ResourceId

The ID of the hosted zone that you want Amazon Route 53 to simulate a query for.

recordName :: Text

The name of the resource record set that you want Amazon Route 53 to simulate a query for.

recordType :: RRType

The type of the resource record set.

newTestDNSAnswer

Arguments

:: ResourceId
$sel:hostedZoneId:TestDNSAnswer'

-> Text
$sel:recordName:TestDNSAnswer'

-> RRType
$sel:recordType:TestDNSAnswer'

-> TestDNSAnswer

Create a value of TestDNSAnswer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolverIP:TestDNSAnswer', testDNSAnswer_resolverIP - If you want to simulate a request from a specific DNS resolver, specify the IP address for that resolver. If you omit this value, TestDnsAnswer uses the IP address of a DNS resolver in the Amazon Web Services US East (N. Virginia) Region (us-east-1).

$sel:eDNS0ClientSubnetIP:TestDNSAnswer', testDNSAnswer_eDNS0ClientSubnetIP - If the resolver that you specified for resolverip supports EDNS0, specify the IPv4 or IPv6 address of a client in the applicable location, for example, 192.0.2.44 or 2001:db8:85a3::8a2e:370:7334.

$sel:eDNS0ClientSubnetMask:TestDNSAnswer', testDNSAnswer_eDNS0ClientSubnetMask - If you specify an IP address for edns0clientsubnetip, you can optionally specify the number of bits of the IP address that you want the checking tool to include in the DNS query. For example, if you specify 192.0.2.44 for edns0clientsubnetip and 24 for edns0clientsubnetmask, the checking tool will simulate a request from 192.0.2.0/24. The default value is 24 bits for IPv4 addresses and 64 bits for IPv6 addresses.

The range of valid values depends on whether edns0clientsubnetip is an IPv4 or an IPv6 address:

IPv4: Specify a value between 0 and 32

IPv6: Specify a value between 0 and 128

$sel:hostedZoneId:TestDNSAnswer', testDNSAnswer_hostedZoneId - The ID of the hosted zone that you want Amazon Route 53 to simulate a query for.

$sel:recordName:TestDNSAnswer', testDNSAnswer_recordName - The name of the resource record set that you want Amazon Route 53 to simulate a query for.

$sel:recordType:TestDNSAnswer', testDNSAnswer_recordType - The type of the resource record set.

Request Lenses

testDNSAnswer_resolverIP :: Lens' TestDNSAnswer (Maybe Text)

If you want to simulate a request from a specific DNS resolver, specify the IP address for that resolver. If you omit this value, TestDnsAnswer uses the IP address of a DNS resolver in the Amazon Web Services US East (N. Virginia) Region (us-east-1).

testDNSAnswer_eDNS0ClientSubnetIP :: Lens' TestDNSAnswer (Maybe Text)

If the resolver that you specified for resolverip supports EDNS0, specify the IPv4 or IPv6 address of a client in the applicable location, for example, 192.0.2.44 or 2001:db8:85a3::8a2e:370:7334.

testDNSAnswer_eDNS0ClientSubnetMask :: Lens' TestDNSAnswer (Maybe Text)

If you specify an IP address for edns0clientsubnetip, you can optionally specify the number of bits of the IP address that you want the checking tool to include in the DNS query. For example, if you specify 192.0.2.44 for edns0clientsubnetip and 24 for edns0clientsubnetmask, the checking tool will simulate a request from 192.0.2.0/24. The default value is 24 bits for IPv4 addresses and 64 bits for IPv6 addresses.

The range of valid values depends on whether edns0clientsubnetip is an IPv4 or an IPv6 address:

IPv4: Specify a value between 0 and 32

IPv6: Specify a value between 0 and 128

testDNSAnswer_hostedZoneId :: Lens' TestDNSAnswer ResourceId

The ID of the hosted zone that you want Amazon Route 53 to simulate a query for.

testDNSAnswer_recordName :: Lens' TestDNSAnswer Text

The name of the resource record set that you want Amazon Route 53 to simulate a query for.

testDNSAnswer_recordType :: Lens' TestDNSAnswer RRType

The type of the resource record set.

Destructuring the Response

data TestDNSAnswerResponse

A complex type that contains the response to a TestDNSAnswer request.

See: newTestDNSAnswerResponse smart constructor.

Constructors

TestDNSAnswerResponse'

Fields

httpStatus :: Int

The response's http status code.

nameserver :: Text

The Amazon Route 53 name server used to respond to the request.

recordName :: Text

The name of the resource record set that you submitted a request for.

recordType :: RRType

The type of the resource record set that you submitted a request for.

recordData :: [Text]

A list that contains values that Amazon Route 53 returned for this resource record set.

responseCode :: Text

A code that indicates whether the request is valid or not. The most common response code is NOERROR, meaning that the request is valid. If the response is not valid, Amazon Route 53 returns a response code that describes the error. For a list of possible response codes, see DNS RCODES on the IANA website.

protocol :: Text

The protocol that Amazon Route 53 used to respond to the request, either UDP or TCP.

newTestDNSAnswerResponse

Arguments

:: Int
$sel:httpStatus:TestDNSAnswerResponse'

-> Text
$sel:nameserver:TestDNSAnswerResponse'

-> Text
$sel:recordName:TestDNSAnswer'

-> RRType
$sel:recordType:TestDNSAnswer'

-> Text
$sel:responseCode:TestDNSAnswerResponse'

-> Text
$sel:protocol:TestDNSAnswerResponse'

-> TestDNSAnswerResponse

Create a value of TestDNSAnswerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TestDNSAnswerResponse', testDNSAnswerResponse_httpStatus - The response's http status code.

$sel:nameserver:TestDNSAnswerResponse', testDNSAnswerResponse_nameserver - The Amazon Route 53 name server used to respond to the request.

$sel:recordName:TestDNSAnswer', testDNSAnswerResponse_recordName - The name of the resource record set that you submitted a request for.

$sel:recordType:TestDNSAnswer', testDNSAnswerResponse_recordType - The type of the resource record set that you submitted a request for.

$sel:recordData:TestDNSAnswerResponse', testDNSAnswerResponse_recordData - A list that contains values that Amazon Route 53 returned for this resource record set.

$sel:responseCode:TestDNSAnswerResponse', testDNSAnswerResponse_responseCode - A code that indicates whether the request is valid or not. The most common response code is NOERROR, meaning that the request is valid. If the response is not valid, Amazon Route 53 returns a response code that describes the error. For a list of possible response codes, see DNS RCODES on the IANA website.

$sel:protocol:TestDNSAnswerResponse', testDNSAnswerResponse_protocol - The protocol that Amazon Route 53 used to respond to the request, either UDP or TCP.

Response Lenses

testDNSAnswerResponse_httpStatus :: Lens' TestDNSAnswerResponse Int

The response's http status code.

testDNSAnswerResponse_nameserver :: Lens' TestDNSAnswerResponse Text

The Amazon Route 53 name server used to respond to the request.

testDNSAnswerResponse_recordName :: Lens' TestDNSAnswerResponse Text

The name of the resource record set that you submitted a request for.

testDNSAnswerResponse_recordType :: Lens' TestDNSAnswerResponse RRType

The type of the resource record set that you submitted a request for.

testDNSAnswerResponse_recordData :: Lens' TestDNSAnswerResponse [Text]

A list that contains values that Amazon Route 53 returned for this resource record set.

testDNSAnswerResponse_responseCode :: Lens' TestDNSAnswerResponse Text

A code that indicates whether the request is valid or not. The most common response code is NOERROR, meaning that the request is valid. If the response is not valid, Amazon Route 53 returns a response code that describes the error. For a list of possible response codes, see DNS RCODES on the IANA website.

testDNSAnswerResponse_protocol :: Lens' TestDNSAnswerResponse Text

The protocol that Amazon Route 53 used to respond to the request, either UDP or TCP.