Network.AWS.Route53.Types.AccountLimit

Description

Documentation

data AccountLimit

A complex type that contains the type of limit that you specified in the request and the current value for that limit.

See: newAccountLimit smart constructor.

Constructors

AccountLimit'

Fields

type' :: AccountLimitType

The limit that you requested. Valid values include the following:

MAX_HEALTH_CHECKS_BY_OWNER: The maximum number of health checks that you can create using the current account.

MAX_HOSTED_ZONES_BY_OWNER: The maximum number of hosted zones that you can create using the current account.

MAX_REUSABLE_DELEGATION_SETS_BY_OWNER: The maximum number of reusable delegation sets that you can create using the current account.

MAX_TRAFFIC_POLICIES_BY_OWNER: The maximum number of traffic policies that you can create using the current account.

MAX_TRAFFIC_POLICY_INSTANCES_BY_OWNER: The maximum number of traffic policy instances that you can create using the current account. (Traffic policy instances are referred to as traffic flow policy records in the Amazon Route 53 console.)

value :: Natural

The current value for the limit that is specified by Type.

newAccountLimit

Arguments

:: AccountLimitType
$sel:type':AccountLimit'

-> Natural
$sel:value:AccountLimit'

-> AccountLimit

Create a value of AccountLimit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':AccountLimit', accountLimit_type - The limit that you requested. Valid values include the following:

MAX_HEALTH_CHECKS_BY_OWNER: The maximum number of health checks that you can create using the current account.

MAX_HOSTED_ZONES_BY_OWNER: The maximum number of hosted zones that you can create using the current account.

MAX_REUSABLE_DELEGATION_SETS_BY_OWNER: The maximum number of reusable delegation sets that you can create using the current account.

MAX_TRAFFIC_POLICIES_BY_OWNER: The maximum number of traffic policies that you can create using the current account.

MAX_TRAFFIC_POLICY_INSTANCES_BY_OWNER: The maximum number of traffic policy instances that you can create using the current account. (Traffic policy instances are referred to as traffic flow policy records in the Amazon Route 53 console.)

$sel:value:AccountLimit', accountLimit_value - The current value for the limit that is specified by Type.

accountLimit_type :: Lens' AccountLimit AccountLimitType

The limit that you requested. Valid values include the following:

MAX_HEALTH_CHECKS_BY_OWNER: The maximum number of health checks that you can create using the current account.

MAX_HOSTED_ZONES_BY_OWNER: The maximum number of hosted zones that you can create using the current account.

MAX_REUSABLE_DELEGATION_SETS_BY_OWNER: The maximum number of reusable delegation sets that you can create using the current account.

MAX_TRAFFIC_POLICIES_BY_OWNER: The maximum number of traffic policies that you can create using the current account.

MAX_TRAFFIC_POLICY_INSTANCES_BY_OWNER: The maximum number of traffic policy instances that you can create using the current account. (Traffic policy instances are referred to as traffic flow policy records in the Amazon Route 53 console.)

accountLimit_value :: Lens' AccountLimit Natural

The current value for the limit that is specified by Type.