Network.AWS.Route53.Types.AlarmIdentifier

Description

Documentation

data AlarmIdentifier

A complex type that identifies the CloudWatch alarm that you want Amazon Route 53 health checkers to use to determine whether the specified health check is healthy.

See: newAlarmIdentifier smart constructor.

Constructors

AlarmIdentifier'

Fields

region :: CloudWatchRegion

For the CloudWatch alarm that you want Route 53 health checkers to use to determine whether this health check is healthy, the region that the alarm was created in.

For the current list of CloudWatch regions, see Amazon CloudWatch endpoints and quotas in the Amazon Web Services General Reference.

name :: Text

The name of the CloudWatch alarm that you want Amazon Route 53 health checkers to use to determine whether this health check is healthy.

Route 53 supports CloudWatch alarms with the following features:

Standard-resolution metrics. High-resolution metrics aren't supported. For more information, see High-Resolution Metrics in the Amazon CloudWatch User Guide.

Statistics: Average, Minimum, Maximum, Sum, and SampleCount. Extended statistics aren't supported.

newAlarmIdentifier

Arguments

:: CloudWatchRegion
$sel:region:AlarmIdentifier'

-> Text
$sel:name:AlarmIdentifier'

-> AlarmIdentifier

Create a value of AlarmIdentifier with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:region:AlarmIdentifier', alarmIdentifier_region - For the CloudWatch alarm that you want Route 53 health checkers to use to determine whether this health check is healthy, the region that the alarm was created in.

For the current list of CloudWatch regions, see Amazon CloudWatch endpoints and quotas in the Amazon Web Services General Reference.

$sel:name:AlarmIdentifier', alarmIdentifier_name - The name of the CloudWatch alarm that you want Amazon Route 53 health checkers to use to determine whether this health check is healthy.

Route 53 supports CloudWatch alarms with the following features:

Standard-resolution metrics. High-resolution metrics aren't supported. For more information, see High-Resolution Metrics in the Amazon CloudWatch User Guide.

Statistics: Average, Minimum, Maximum, Sum, and SampleCount. Extended statistics aren't supported.

alarmIdentifier_region :: Lens' AlarmIdentifier CloudWatchRegion

For the CloudWatch alarm that you want Route 53 health checkers to use to determine whether this health check is healthy, the region that the alarm was created in.

For the current list of CloudWatch regions, see Amazon CloudWatch endpoints and quotas in the Amazon Web Services General Reference.

alarmIdentifier_name :: Lens' AlarmIdentifier Text

The name of the CloudWatch alarm that you want Amazon Route 53 health checkers to use to determine whether this health check is healthy.

Route 53 supports CloudWatch alarms with the following features:

Standard-resolution metrics. High-resolution metrics aren't supported. For more information, see High-Resolution Metrics in the Amazon CloudWatch User Guide.

Statistics: Average, Minimum, Maximum, Sum, and SampleCount. Extended statistics aren't supported.