Network.AWS.Route53.Types.ChangeBatch

Description

Documentation

data ChangeBatch

The information for a change request.

See: newChangeBatch smart constructor.

Constructors

ChangeBatch'

Fields

comment :: Maybe Text

Optional: Any comments you want to include about a change batch request.

changes :: NonEmpty Change

Information about the changes to make to the record sets.

newChangeBatch

Arguments

:: NonEmpty Change
$sel:changes:ChangeBatch'

-> ChangeBatch

Create a value of ChangeBatch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:comment:ChangeBatch', changeBatch_comment - Optional: Any comments you want to include about a change batch request.

$sel:changes:ChangeBatch', changeBatch_changes - Information about the changes to make to the record sets.

changeBatch_comment :: Lens' ChangeBatch (Maybe Text)

Optional: Any comments you want to include about a change batch request.

changeBatch_changes :: Lens' ChangeBatch (NonEmpty Change)

Information about the changes to make to the record sets.