Network.AWS.Route53.Types.CloudWatchAlarmConfiguration

Description

Documentation

data CloudWatchAlarmConfiguration

A complex type that contains information about the CloudWatch alarm that Amazon Route 53 is monitoring for this health check.

See: newCloudWatchAlarmConfiguration smart constructor.

Constructors

CloudWatchAlarmConfiguration'

Fields

dimensions :: Maybe [Dimension]

For the metric that the CloudWatch alarm is associated with, a complex type that contains information about the dimensions for the metric. For information, see Amazon CloudWatch Namespaces, Dimensions, and Metrics Reference in the Amazon CloudWatch User Guide.

evaluationPeriods :: Natural

For the metric that the CloudWatch alarm is associated with, the number of periods that the metric is compared to the threshold.

threshold :: Double

For the metric that the CloudWatch alarm is associated with, the value the metric is compared with.

comparisonOperator :: ComparisonOperator

For the metric that the CloudWatch alarm is associated with, the arithmetic operation that is used for the comparison.

period :: Natural

For the metric that the CloudWatch alarm is associated with, the duration of one evaluation period in seconds.

metricName :: Text

The name of the CloudWatch metric that the alarm is associated with.

namespace :: Text

The namespace of the metric that the alarm is associated with. For more information, see Amazon CloudWatch Namespaces, Dimensions, and Metrics Reference in the Amazon CloudWatch User Guide.

statistic :: Statistic

For the metric that the CloudWatch alarm is associated with, the statistic that is applied to the metric.

newCloudWatchAlarmConfiguration

Arguments

:: Natural
$sel:evaluationPeriods:CloudWatchAlarmConfiguration'

-> Double
$sel:threshold:CloudWatchAlarmConfiguration'

-> ComparisonOperator
$sel:comparisonOperator:CloudWatchAlarmConfiguration'

-> Natural
$sel:period:CloudWatchAlarmConfiguration'

-> Text
$sel:metricName:CloudWatchAlarmConfiguration'

-> Text
$sel:namespace:CloudWatchAlarmConfiguration'

-> Statistic
$sel:statistic:CloudWatchAlarmConfiguration'

-> CloudWatchAlarmConfiguration

Create a value of CloudWatchAlarmConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimensions:CloudWatchAlarmConfiguration', cloudWatchAlarmConfiguration_dimensions - For the metric that the CloudWatch alarm is associated with, a complex type that contains information about the dimensions for the metric. For information, see Amazon CloudWatch Namespaces, Dimensions, and Metrics Reference in the Amazon CloudWatch User Guide.

$sel:evaluationPeriods:CloudWatchAlarmConfiguration', cloudWatchAlarmConfiguration_evaluationPeriods - For the metric that the CloudWatch alarm is associated with, the number of periods that the metric is compared to the threshold.

$sel:threshold:CloudWatchAlarmConfiguration', cloudWatchAlarmConfiguration_threshold - For the metric that the CloudWatch alarm is associated with, the value the metric is compared with.

$sel:comparisonOperator:CloudWatchAlarmConfiguration', cloudWatchAlarmConfiguration_comparisonOperator - For the metric that the CloudWatch alarm is associated with, the arithmetic operation that is used for the comparison.

$sel:period:CloudWatchAlarmConfiguration', cloudWatchAlarmConfiguration_period - For the metric that the CloudWatch alarm is associated with, the duration of one evaluation period in seconds.

$sel:metricName:CloudWatchAlarmConfiguration', cloudWatchAlarmConfiguration_metricName - The name of the CloudWatch metric that the alarm is associated with.

$sel:namespace:CloudWatchAlarmConfiguration', cloudWatchAlarmConfiguration_namespace - The namespace of the metric that the alarm is associated with. For more information, see Amazon CloudWatch Namespaces, Dimensions, and Metrics Reference in the Amazon CloudWatch User Guide.

$sel:statistic:CloudWatchAlarmConfiguration', cloudWatchAlarmConfiguration_statistic - For the metric that the CloudWatch alarm is associated with, the statistic that is applied to the metric.

cloudWatchAlarmConfiguration_dimensions :: Lens' CloudWatchAlarmConfiguration (Maybe [Dimension])

For the metric that the CloudWatch alarm is associated with, a complex type that contains information about the dimensions for the metric. For information, see Amazon CloudWatch Namespaces, Dimensions, and Metrics Reference in the Amazon CloudWatch User Guide.

cloudWatchAlarmConfiguration_evaluationPeriods :: Lens' CloudWatchAlarmConfiguration Natural

For the metric that the CloudWatch alarm is associated with, the number of periods that the metric is compared to the threshold.

cloudWatchAlarmConfiguration_threshold :: Lens' CloudWatchAlarmConfiguration Double

For the metric that the CloudWatch alarm is associated with, the value the metric is compared with.

cloudWatchAlarmConfiguration_comparisonOperator :: Lens' CloudWatchAlarmConfiguration ComparisonOperator

For the metric that the CloudWatch alarm is associated with, the arithmetic operation that is used for the comparison.

cloudWatchAlarmConfiguration_period :: Lens' CloudWatchAlarmConfiguration Natural

For the metric that the CloudWatch alarm is associated with, the duration of one evaluation period in seconds.

cloudWatchAlarmConfiguration_metricName :: Lens' CloudWatchAlarmConfiguration Text

The name of the CloudWatch metric that the alarm is associated with.

cloudWatchAlarmConfiguration_namespace :: Lens' CloudWatchAlarmConfiguration Text

The namespace of the metric that the alarm is associated with. For more information, see Amazon CloudWatch Namespaces, Dimensions, and Metrics Reference in the Amazon CloudWatch User Guide.

cloudWatchAlarmConfiguration_statistic :: Lens' CloudWatchAlarmConfiguration Statistic

For the metric that the CloudWatch alarm is associated with, the statistic that is applied to the metric.