Network.AWS.Route53.Types.DNSSECStatus

Description

Documentation

data DNSSECStatus

A string repesenting the status of DNSSEC signing.

See: newDNSSECStatus smart constructor.

Constructors

DNSSECStatus'

Fields

statusMessage :: Maybe Text

The status message provided for the following DNSSEC signing status: INTERNAL_FAILURE. The status message includes information about what the problem might be and steps that you can take to correct the issue.

serveSignature :: Maybe Text

A string that represents the current hosted zone signing status.

Status can have one of the following values:

SIGNING

DNSSEC signing is enabled for the hosted zone.

NOT_SIGNING

DNSSEC signing is not enabled for the hosted zone.

DELETING

DNSSEC signing is in the process of being removed for the hosted zone.

ACTION_NEEDED

There is a problem with signing in the hosted zone that requires you to take action to resolve. For example, the customer managed customer master key (CMK) might have been deleted, or the permissions for the customer managed CMK might have been changed.

INTERNAL_FAILURE

There was an error during a request. Before you can continue to work with DNSSEC signing, including with key-signing keys (KSKs), you must correct the problem by enabling or disabling DNSSEC signing for the hosted zone.

newDNSSECStatus :: DNSSECStatus

Create a value of DNSSECStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusMessage:DNSSECStatus', dNSSECStatus_statusMessage - The status message provided for the following DNSSEC signing status: INTERNAL_FAILURE. The status message includes information about what the problem might be and steps that you can take to correct the issue.

$sel:serveSignature:DNSSECStatus', dNSSECStatus_serveSignature - A string that represents the current hosted zone signing status.

Status can have one of the following values:

SIGNING

DNSSEC signing is enabled for the hosted zone.

NOT_SIGNING

DNSSEC signing is not enabled for the hosted zone.

DELETING

DNSSEC signing is in the process of being removed for the hosted zone.

ACTION_NEEDED

There is a problem with signing in the hosted zone that requires you to take action to resolve. For example, the customer managed customer master key (CMK) might have been deleted, or the permissions for the customer managed CMK might have been changed.

INTERNAL_FAILURE

There was an error during a request. Before you can continue to work with DNSSEC signing, including with key-signing keys (KSKs), you must correct the problem by enabling or disabling DNSSEC signing for the hosted zone.

dNSSECStatus_statusMessage :: Lens' DNSSECStatus (Maybe Text)

The status message provided for the following DNSSEC signing status: INTERNAL_FAILURE. The status message includes information about what the problem might be and steps that you can take to correct the issue.

dNSSECStatus_serveSignature :: Lens' DNSSECStatus (Maybe Text)

A string that represents the current hosted zone signing status.

Status can have one of the following values:

SIGNING

DNSSEC signing is enabled for the hosted zone.

NOT_SIGNING

DNSSEC signing is not enabled for the hosted zone.

DELETING

DNSSEC signing is in the process of being removed for the hosted zone.

ACTION_NEEDED

There is a problem with signing in the hosted zone that requires you to take action to resolve. For example, the customer managed customer master key (CMK) might have been deleted, or the permissions for the customer managed CMK might have been changed.

INTERNAL_FAILURE

There was an error during a request. Before you can continue to work with DNSSEC signing, including with key-signing keys (KSKs), you must correct the problem by enabling or disabling DNSSEC signing for the hosted zone.