Network.AWS.Route53.Types.GeoLocation

Description

Documentation

data GeoLocation

A complex type that contains information about a geographic location.

See: newGeoLocation smart constructor.

Constructors

GeoLocation'

Fields

continentCode :: Maybe Text

The two-letter code for the continent.

Amazon Route 53 supports the following continent codes:

AF: Africa

AN: Antarctica

AS: Asia

EU: Europe

OC: Oceania

NA: North America

SA: South America

Constraint: Specifying ContinentCode with either CountryCode or SubdivisionCode returns an InvalidInput error.

subdivisionCode :: Maybe Text

For geolocation resource record sets, the two-letter code for a state of the United States. Route 53 doesn't support any other values for SubdivisionCode. For a list of state abbreviations, see Appendix B: Two–Letter State and Possession Abbreviations on the United States Postal Service website.

If you specify subdivisioncode, you must also specify US for CountryCode.

countryCode :: Maybe Text

For geolocation resource record sets, the two-letter code for a country.

Amazon Route 53 uses the two-letter country codes that are specified in ISO standard 3166-1 alpha-2.

newGeoLocation :: GeoLocation

Create a value of GeoLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:continentCode:GeoLocation', geoLocation_continentCode - The two-letter code for the continent.

Amazon Route 53 supports the following continent codes:

AF: Africa

AN: Antarctica

AS: Asia

EU: Europe

OC: Oceania

NA: North America

SA: South America

Constraint: Specifying ContinentCode with either CountryCode or SubdivisionCode returns an InvalidInput error.

$sel:subdivisionCode:GeoLocation', geoLocation_subdivisionCode - For geolocation resource record sets, the two-letter code for a state of the United States. Route 53 doesn't support any other values for SubdivisionCode. For a list of state abbreviations, see Appendix B: Two–Letter State and Possession Abbreviations on the United States Postal Service website.

If you specify subdivisioncode, you must also specify US for CountryCode.

$sel:countryCode:GeoLocation', geoLocation_countryCode - For geolocation resource record sets, the two-letter code for a country.

Amazon Route 53 uses the two-letter country codes that are specified in ISO standard 3166-1 alpha-2.

geoLocation_continentCode :: Lens' GeoLocation (Maybe Text)

The two-letter code for the continent.

Amazon Route 53 supports the following continent codes:

AF: Africa

AN: Antarctica

AS: Asia

EU: Europe

OC: Oceania

NA: North America

SA: South America

Constraint: Specifying ContinentCode with either CountryCode or SubdivisionCode returns an InvalidInput error.

geoLocation_subdivisionCode :: Lens' GeoLocation (Maybe Text)

For geolocation resource record sets, the two-letter code for a state of the United States. Route 53 doesn't support any other values for SubdivisionCode. For a list of state abbreviations, see Appendix B: Two–Letter State and Possession Abbreviations on the United States Postal Service website.

If you specify subdivisioncode, you must also specify US for CountryCode.

geoLocation_countryCode :: Lens' GeoLocation (Maybe Text)

For geolocation resource record sets, the two-letter code for a country.

Amazon Route 53 uses the two-letter country codes that are specified in ISO standard 3166-1 alpha-2.