Network.AWS.Route53.Types.HostedZoneSummary

Description

Documentation

data HostedZoneSummary

In the response to a ListHostedZonesByVPC request, the HostedZoneSummaries element contains one HostedZoneSummary element for each hosted zone that the specified Amazon VPC is associated with. Each HostedZoneSummary element contains the hosted zone name and ID, and information about who owns the hosted zone.

See: newHostedZoneSummary smart constructor.

Constructors

HostedZoneSummary'

Fields

hostedZoneId :: ResourceId

The Route 53 hosted zone ID of a private hosted zone that the specified VPC is associated with.

name :: Text

The name of the private hosted zone, such as example.com.

owner :: HostedZoneOwner

The owner of a private hosted zone that the specified VPC is associated with. The owner can be either an Amazon Web Services account or an Amazon Web Services service.

newHostedZoneSummary

Arguments

:: ResourceId
$sel:hostedZoneId:HostedZoneSummary'

-> Text
$sel:name:HostedZoneSummary'

-> HostedZoneOwner
$sel:owner:HostedZoneSummary'

-> HostedZoneSummary

Create a value of HostedZoneSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostedZoneId:HostedZoneSummary', hostedZoneSummary_hostedZoneId - The Route 53 hosted zone ID of a private hosted zone that the specified VPC is associated with.

$sel:name:HostedZoneSummary', hostedZoneSummary_name - The name of the private hosted zone, such as example.com.

$sel:owner:HostedZoneSummary', hostedZoneSummary_owner - The owner of a private hosted zone that the specified VPC is associated with. The owner can be either an Amazon Web Services account or an Amazon Web Services service.

hostedZoneSummary_hostedZoneId :: Lens' HostedZoneSummary ResourceId

The Route 53 hosted zone ID of a private hosted zone that the specified VPC is associated with.

hostedZoneSummary_name :: Lens' HostedZoneSummary Text

The name of the private hosted zone, such as example.com.

hostedZoneSummary_owner :: Lens' HostedZoneSummary HostedZoneOwner

The owner of a private hosted zone that the specified VPC is associated with. The owner can be either an Amazon Web Services account or an Amazon Web Services service.