Network.AWS.Route53.Types.KeySigningKey

Description

Documentation

data KeySigningKey

A key-signing key (KSK) is a complex type that represents a public/private key pair. The private key is used to generate a digital signature for the zone signing key (ZSK). The public key is stored in the DNS and is used to authenticate the ZSK. A KSK is always associated with a hosted zone; it cannot exist by itself.

See: newKeySigningKey smart constructor.

Constructors

KeySigningKey'

Fields

digestAlgorithmType :: Maybe Int

An integer used to represent the delegation signer digest algorithm. This value must follow the guidelines provided by RFC-8624 Section 3.3.

lastModifiedDate :: Maybe ISO8601

The last time that the key-signing key (KSK) was changed.

statusMessage :: Maybe Text

The status message provided for the following key-signing key (KSK) statuses: ACTION_NEEDED or INTERNAL_FAILURE. The status message includes information about what the problem might be and steps that you can take to correct the issue.

createdDate :: Maybe ISO8601

The date when the key-signing key (KSK) was created.

signingAlgorithmMnemonic :: Maybe Text

A string used to represent the signing algorithm. This value must follow the guidelines provided by RFC-8624 Section 3.1.

status :: Maybe Text

A string that represents the current key-signing key (KSK) status.

Status can have one of the following values:

ACTIVE

The KSK is being used for signing.

INACTIVE

The KSK is not being used for signing.

DELETING

The KSK is in the process of being deleted.

ACTION_NEEDED

There is a problem with the KSK that requires you to take action to resolve. For example, the customer managed customer master key (CMK) might have been deleted, or the permissions for the customer managed CMK might have been changed.

INTERNAL_FAILURE

There was an error during a request. Before you can continue to work with DNSSEC signing, including actions that involve this KSK, you must correct the problem. For example, you may need to activate or deactivate the KSK.

publicKey :: Maybe Text

The public key, represented as a Base64 encoding, as required by RFC-4034 Page 5.

digestValue :: Maybe Text

A cryptographic digest of a DNSKEY resource record (RR). DNSKEY records are used to publish the public key that resolvers can use to verify DNSSEC signatures that are used to secure certain kinds of information provided by the DNS system.

dNSKEYRecord :: Maybe Text

A string that represents a DNSKEY record.

name :: Maybe Text

A string used to identify a key-signing key (KSK). Name can include numbers, letters, and underscores (_). Name must be unique for each key-signing key in the same hosted zone.

digestAlgorithmMnemonic :: Maybe Text

A string used to represent the delegation signer digest algorithm. This value must follow the guidelines provided by RFC-8624 Section 3.3.

signingAlgorithmType :: Maybe Int

An integer used to represent the signing algorithm. This value must follow the guidelines provided by RFC-8624 Section 3.1.

flag :: Maybe Int

An integer that specifies how the key is used. For key-signing key (KSK), this value is always 257.

kmsArn :: Maybe Text

The Amazon resource name (ARN) used to identify the customer managed customer master key (CMK) in Key Management Service (KMS). The KmsArn must be unique for each key-signing key (KSK) in a single hosted zone.

You must configure the CMK as follows:

Status

Enabled

Key spec

ECC_NIST_P256

Key usage

Sign and verify

Key policy

The key policy must give permission for the following actions:

DescribeKey

GetPublicKey

Sign

The key policy must also include the Amazon Route 53 service in the principal for your account. Specify the following:

"Service": "dnssec-route53.amazonaws.com"

For more information about working with the customer managed CMK in KMS, see Key Management Service concepts.

keyTag :: Maybe Natural

An integer used to identify the DNSSEC record for the domain name. The process used to calculate the value is described in RFC-4034 Appendix B.

dSRecord :: Maybe Text

A string that represents a delegation signer (DS) record.

newKeySigningKey :: KeySigningKey

Create a value of KeySigningKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:digestAlgorithmType:KeySigningKey', keySigningKey_digestAlgorithmType - An integer used to represent the delegation signer digest algorithm. This value must follow the guidelines provided by RFC-8624 Section 3.3.

$sel:lastModifiedDate:KeySigningKey', keySigningKey_lastModifiedDate - The last time that the key-signing key (KSK) was changed.

$sel:statusMessage:KeySigningKey', keySigningKey_statusMessage - The status message provided for the following key-signing key (KSK) statuses: ACTION_NEEDED or INTERNAL_FAILURE. The status message includes information about what the problem might be and steps that you can take to correct the issue.

$sel:createdDate:KeySigningKey', keySigningKey_createdDate - The date when the key-signing key (KSK) was created.

$sel:signingAlgorithmMnemonic:KeySigningKey', keySigningKey_signingAlgorithmMnemonic - A string used to represent the signing algorithm. This value must follow the guidelines provided by RFC-8624 Section 3.1.

$sel:status:KeySigningKey', keySigningKey_status - A string that represents the current key-signing key (KSK) status.

Status can have one of the following values:

ACTIVE

The KSK is being used for signing.

INACTIVE

The KSK is not being used for signing.

DELETING

The KSK is in the process of being deleted.

ACTION_NEEDED

There is a problem with the KSK that requires you to take action to resolve. For example, the customer managed customer master key (CMK) might have been deleted, or the permissions for the customer managed CMK might have been changed.

INTERNAL_FAILURE

There was an error during a request. Before you can continue to work with DNSSEC signing, including actions that involve this KSK, you must correct the problem. For example, you may need to activate or deactivate the KSK.

$sel:publicKey:KeySigningKey', keySigningKey_publicKey - The public key, represented as a Base64 encoding, as required by RFC-4034 Page 5.

$sel:digestValue:KeySigningKey', keySigningKey_digestValue - A cryptographic digest of a DNSKEY resource record (RR). DNSKEY records are used to publish the public key that resolvers can use to verify DNSSEC signatures that are used to secure certain kinds of information provided by the DNS system.

$sel:dNSKEYRecord:KeySigningKey', keySigningKey_dNSKEYRecord - A string that represents a DNSKEY record.

$sel:name:KeySigningKey', keySigningKey_name - A string used to identify a key-signing key (KSK). Name can include numbers, letters, and underscores (_). Name must be unique for each key-signing key in the same hosted zone.

$sel:digestAlgorithmMnemonic:KeySigningKey', keySigningKey_digestAlgorithmMnemonic - A string used to represent the delegation signer digest algorithm. This value must follow the guidelines provided by RFC-8624 Section 3.3.

$sel:signingAlgorithmType:KeySigningKey', keySigningKey_signingAlgorithmType - An integer used to represent the signing algorithm. This value must follow the guidelines provided by RFC-8624 Section 3.1.

$sel:flag:KeySigningKey', keySigningKey_flag - An integer that specifies how the key is used. For key-signing key (KSK), this value is always 257.

$sel:kmsArn:KeySigningKey', keySigningKey_kmsArn - The Amazon resource name (ARN) used to identify the customer managed customer master key (CMK) in Key Management Service (KMS). The KmsArn must be unique for each key-signing key (KSK) in a single hosted zone.

You must configure the CMK as follows:

Status

Enabled

Key spec

ECC_NIST_P256

Key usage

Sign and verify

Key policy

The key policy must give permission for the following actions:

DescribeKey

GetPublicKey

Sign

The key policy must also include the Amazon Route 53 service in the principal for your account. Specify the following:

"Service": "dnssec-route53.amazonaws.com"

For more information about working with the customer managed CMK in KMS, see Key Management Service concepts.

$sel:keyTag:KeySigningKey', keySigningKey_keyTag - An integer used to identify the DNSSEC record for the domain name. The process used to calculate the value is described in RFC-4034 Appendix B.

$sel:dSRecord:KeySigningKey', keySigningKey_dSRecord - A string that represents a delegation signer (DS) record.

keySigningKey_digestAlgorithmType :: Lens' KeySigningKey (Maybe Int)

An integer used to represent the delegation signer digest algorithm. This value must follow the guidelines provided by RFC-8624 Section 3.3.

keySigningKey_lastModifiedDate :: Lens' KeySigningKey (Maybe UTCTime)

The last time that the key-signing key (KSK) was changed.

keySigningKey_statusMessage :: Lens' KeySigningKey (Maybe Text)

The status message provided for the following key-signing key (KSK) statuses: ACTION_NEEDED or INTERNAL_FAILURE. The status message includes information about what the problem might be and steps that you can take to correct the issue.

keySigningKey_createdDate :: Lens' KeySigningKey (Maybe UTCTime)

The date when the key-signing key (KSK) was created.

keySigningKey_signingAlgorithmMnemonic :: Lens' KeySigningKey (Maybe Text)

A string used to represent the signing algorithm. This value must follow the guidelines provided by RFC-8624 Section 3.1.

keySigningKey_status :: Lens' KeySigningKey (Maybe Text)

A string that represents the current key-signing key (KSK) status.

Status can have one of the following values:

ACTIVE

The KSK is being used for signing.

INACTIVE

The KSK is not being used for signing.

DELETING

The KSK is in the process of being deleted.

ACTION_NEEDED

There is a problem with the KSK that requires you to take action to resolve. For example, the customer managed customer master key (CMK) might have been deleted, or the permissions for the customer managed CMK might have been changed.

INTERNAL_FAILURE

There was an error during a request. Before you can continue to work with DNSSEC signing, including actions that involve this KSK, you must correct the problem. For example, you may need to activate or deactivate the KSK.

keySigningKey_publicKey :: Lens' KeySigningKey (Maybe Text)

The public key, represented as a Base64 encoding, as required by RFC-4034 Page 5.

keySigningKey_digestValue :: Lens' KeySigningKey (Maybe Text)

A cryptographic digest of a DNSKEY resource record (RR). DNSKEY records are used to publish the public key that resolvers can use to verify DNSSEC signatures that are used to secure certain kinds of information provided by the DNS system.

keySigningKey_dNSKEYRecord :: Lens' KeySigningKey (Maybe Text)

A string that represents a DNSKEY record.

keySigningKey_name :: Lens' KeySigningKey (Maybe Text)

A string used to identify a key-signing key (KSK). Name can include numbers, letters, and underscores (_). Name must be unique for each key-signing key in the same hosted zone.

keySigningKey_digestAlgorithmMnemonic :: Lens' KeySigningKey (Maybe Text)

A string used to represent the delegation signer digest algorithm. This value must follow the guidelines provided by RFC-8624 Section 3.3.

keySigningKey_signingAlgorithmType :: Lens' KeySigningKey (Maybe Int)

An integer used to represent the signing algorithm. This value must follow the guidelines provided by RFC-8624 Section 3.1.

keySigningKey_flag :: Lens' KeySigningKey (Maybe Int)

An integer that specifies how the key is used. For key-signing key (KSK), this value is always 257.

keySigningKey_kmsArn :: Lens' KeySigningKey (Maybe Text)

The Amazon resource name (ARN) used to identify the customer managed customer master key (CMK) in Key Management Service (KMS). The KmsArn must be unique for each key-signing key (KSK) in a single hosted zone.

You must configure the CMK as follows:

Status

Enabled

Key spec

ECC_NIST_P256

Key usage

Sign and verify

Key policy

The key policy must give permission for the following actions:

DescribeKey

GetPublicKey

Sign

The key policy must also include the Amazon Route 53 service in the principal for your account. Specify the following:

"Service": "dnssec-route53.amazonaws.com"

For more information about working with the customer managed CMK in KMS, see Key Management Service concepts.

keySigningKey_keyTag :: Lens' KeySigningKey (Maybe Natural)

An integer used to identify the DNSSEC record for the domain name. The process used to calculate the value is described in RFC-4034 Appendix B.

keySigningKey_dSRecord :: Lens' KeySigningKey (Maybe Text)

A string that represents a delegation signer (DS) record.