Network.AWS.Route53.UpdateHostedZoneComment

Description

Updates the comment for a specified hosted zone.

Creating a Request

data UpdateHostedZoneComment

A request to update the comment for a hosted zone.

See: newUpdateHostedZoneComment smart constructor.

Constructors

UpdateHostedZoneComment'

Fields

comment :: Maybe Text

The new comment for the hosted zone. If you don't specify a value for Comment, Amazon Route 53 deletes the existing value of the Comment element, if any.

id :: ResourceId

The ID for the hosted zone that you want to update the comment for.

newUpdateHostedZoneComment

Arguments

:: ResourceId
$sel:id:UpdateHostedZoneComment'

-> UpdateHostedZoneComment

Create a value of UpdateHostedZoneComment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:comment:UpdateHostedZoneComment', updateHostedZoneComment_comment - The new comment for the hosted zone. If you don't specify a value for Comment, Amazon Route 53 deletes the existing value of the Comment element, if any.

$sel:id:UpdateHostedZoneComment', updateHostedZoneComment_id - The ID for the hosted zone that you want to update the comment for.

Request Lenses

updateHostedZoneComment_comment :: Lens' UpdateHostedZoneComment (Maybe Text)

The new comment for the hosted zone. If you don't specify a value for Comment, Amazon Route 53 deletes the existing value of the Comment element, if any.

updateHostedZoneComment_id :: Lens' UpdateHostedZoneComment ResourceId

The ID for the hosted zone that you want to update the comment for.

Destructuring the Response

data UpdateHostedZoneCommentResponse

A complex type that contains the response to the UpdateHostedZoneComment request.

See: newUpdateHostedZoneCommentResponse smart constructor.

Constructors

UpdateHostedZoneCommentResponse'

Fields

httpStatus :: Int

The response's http status code.

hostedZone :: HostedZone

A complex type that contains the response to the UpdateHostedZoneComment request.

newUpdateHostedZoneCommentResponse

Arguments

:: Int
$sel:httpStatus:UpdateHostedZoneCommentResponse'

-> HostedZone
$sel:hostedZone:UpdateHostedZoneCommentResponse'

-> UpdateHostedZoneCommentResponse

Create a value of UpdateHostedZoneCommentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateHostedZoneCommentResponse', updateHostedZoneCommentResponse_httpStatus - The response's http status code.

$sel:hostedZone:UpdateHostedZoneCommentResponse', updateHostedZoneCommentResponse_hostedZone - A complex type that contains the response to the UpdateHostedZoneComment request.

Response Lenses

updateHostedZoneCommentResponse_httpStatus :: Lens' UpdateHostedZoneCommentResponse Int

The response's http status code.

updateHostedZoneCommentResponse_hostedZone :: Lens' UpdateHostedZoneCommentResponse HostedZone

A complex type that contains the response to the UpdateHostedZoneComment request.