Network.AWS.Route53.V2013_04_01

Description

Amazon Route 53 is a scalable Domain Name System (DNS) web service. It provides secure and reliable routing to your infrastructure that uses Amazon Web Services (AWS) products, such as Amazon Elastic Compute Cloud (Amazon EC2), Elastic Load Balancing, or Amazon Simple Storage Service (Amazon S3). You can also use Amazon Route 53 to route users to your infrastructure outside of AWS.

Documentation

module Network.AWS.Route53.V2013_04_01.ChangeResourceRecordSets

module Network.AWS.Route53.V2013_04_01.ChangeTagsForResource

module Network.AWS.Route53.V2013_04_01.CreateHealthCheck

module Network.AWS.Route53.V2013_04_01.CreateHostedZone

module Network.AWS.Route53.V2013_04_01.DeleteHealthCheck

module Network.AWS.Route53.V2013_04_01.DeleteHostedZone

module Network.AWS.Route53.V2013_04_01.GetChange

module Network.AWS.Route53.V2013_04_01.GetCheckerIpRanges

module Network.AWS.Route53.V2013_04_01.GetHealthCheck

module Network.AWS.Route53.V2013_04_01.GetHealthCheckCount

module Network.AWS.Route53.V2013_04_01.GetHostedZone

module Network.AWS.Route53.V2013_04_01.ListHealthChecks

module Network.AWS.Route53.V2013_04_01.ListHostedZones

module Network.AWS.Route53.V2013_04_01.ListResourceRecordSets

module Network.AWS.Route53.V2013_04_01.ListTagsForResource

module Network.AWS.Route53.V2013_04_01.ListTagsForResources

module Network.AWS.Route53.V2013_04_01.Types

module Network.AWS.Route53.V2013_04_01.UpdateHealthCheck