Network.AWS.Route53.V2013_04_01.CreateHealthCheck

Description

This action creates a new health check. To create a new health check, send a POST request to the 2013-04-01/healthcheck resource. The request body must include an XML document with a CreateHealthCheckRequest element. The response returns the CreateHealthCheckResponse element that contains metadata about the health check.

Documentation

data CreateHealthCheck

Constructors

CreateHealthCheck

Fields

_chcrHealthCheckConfig :: HealthCheckConfig

A complex type that contains health check configuration.

_chcrCallerReference :: Text

A unique string that identifies the request and that allows failed CreateHealthCheck requests to be retried without the risk of executing the operation twice. You must use a unique CallerReference string every time you create a health check. CallerReference can be any unique string; you might choose to use a string that identifies your project. Valid characters are any Unicode code points that are legal in an XML 1.0 document. The UTF-8 encoding of the value must be less than 128 bytes.

chcrHealthCheckConfig :: Lens' CreateHealthCheck HealthCheckConfig

chcrCallerReference :: Lens' CreateHealthCheck Text

data CreateHealthCheckResponse

Constructors

CreateHealthCheckResponse

Fields

_chcsHealthCheck :: HealthCheck

A complex type that contains identifying information about the health check.

_chcsLocation :: Text

The unique URL representing the new health check.

chcsLocation :: Lens' CreateHealthCheckResponse Text

chcsHealthCheck :: Lens' CreateHealthCheckResponse HealthCheck