Network.AWS.Route53.V2013_04_01.DeleteHostedZone

Description

This action deletes a hosted zone. To delete a hosted zone, send a DELETE request to the 2013-04-01hostedzonehosted zone ID resource. For more information about deleting a hosted zone, see Deleting a Hosted Zone in the Amazon Route 53 Developer Guide. You can delete a hosted zone only if there are no resource record sets other than the default SOA record and NS resource record sets. If your hosted zone contains other resource record sets, you must delete them before you can delete your hosted zone. If you try to delete a hosted zone that contains other resource record sets, Route 53 will deny your request with a HostedZoneNotEmpty error. For information about deleting records from your hosted zone, see ChangeResourceRecordSets.

Documentation

data DeleteHostedZone

Constructors

DeleteHostedZone

Fields

_dhzrId :: Text

The ID of the request. Include this ID in a call to GetChange to track when the change has propagated to all Route 53 DNS servers.

dhzrId :: Iso' DeleteHostedZone Text

data DeleteHostedZoneResponse

Constructors

DeleteHostedZoneResponse

Fields

_dhzsChangeInfo :: ChangeInfo

A complex type that contains the ID, the status, and the date and time of your delete request.

dhzsChangeInfo :: Iso' DeleteHostedZoneResponse ChangeInfo