Network.AWS.Route53.V2013_04_01.GetHealthCheckCount

Description

To retrieve a count of all your health checks, send a GET request to the 2013-04-01/healthcheckcount resource.

Documentation

data GetHealthCheckCount

Constructors

GetHealthCheckCount

data GetHealthCheckCountResponse

Constructors

GetHealthCheckCountResponse

Fields

_ghccsHealthCheckCount :: Integer

The number of health checks associated with the current AWS account.

ghccsHealthCheckCount :: Iso' GetHealthCheckCountResponse Integer