Network.AWS.Route53.V2013_04_01.ListHealthChecks

Description

To retrieve a list of your health checks, send a GET request to the 2013-04-01/healthcheck resource. The response to this request includes a HealthChecks element with zero, one, or multiple HealthCheck child elements. By default, the list of health checks is displayed on a single page. You can control the length of the page that is displayed by using the MaxItems parameter. You can use the Marker parameter to control the health check that the list begins with. Amazon Route 53 returns a maximum of 100 items. If you set MaxItems to a value greater than 100, Amazon Route 53 returns only the first 100.

Documentation

listHealthChecks :: ListHealthChecks

Minimum specification for a ListHealthChecks request.

data ListHealthChecks

Constructors

ListHealthChecks

Fields

_lhcrMarker :: Maybe Text

If the request returned more than one page of results, submit another request and specify the value of NextMarker from the last response in the marker parameter to get the next page of results.

_lhcrMaxItems :: Maybe Text

Specify the maximum number of health checks to return per page of results.

lhcrMaxItems :: Lens' ListHealthChecks (Maybe Text)

lhcrMarker :: Lens' ListHealthChecks (Maybe Text)

data ListHealthChecksResponse

Constructors

ListHealthChecksResponse

Fields

_lhcsHealthChecks :: [HealthCheck]

A complex type that contains information about the health checks associated with the current AWS account.

_lhcsMarker :: Text

If the request returned more than one page of results, submit another request and specify the value of NextMarker from the last response in the marker parameter to get the next page of results.

_lhcsMaxItems :: Text

The maximum number of health checks to be included in the response body. If the number of health checks associated with this AWS account exceeds MaxItems, the value of ListHealthChecksResponse$IsTruncated in the response is true. Call ListHealthChecks again and specify the value of ListHealthChecksResponse$NextMarker in the ListHostedZonesRequest$Marker element to get the next page of results.

_lhcsIsTruncated :: Bool

A flag indicating whether there are more health checks to be listed. If your results were truncated, you can make a follow-up request for the next page of results by using the Marker element. Valid Values: true | false.

_lhcsNextMarker :: Maybe Text

Indicates where to continue listing health checks. If ListHealthChecksResponse$IsTruncated is true, make another request to ListHealthChecks and include the value of the NextMarker element in the Marker element to get the next page of results.

lhcsNextMarker :: Lens' ListHealthChecksResponse (Maybe Text)

lhcsMaxItems :: Lens' ListHealthChecksResponse Text

lhcsMarker :: Lens' ListHealthChecksResponse Text

lhcsIsTruncated :: Lens' ListHealthChecksResponse Bool

lhcsHealthChecks :: Lens' ListHealthChecksResponse [HealthCheck]