Network.AWS.Route53AutoNaming.CreateHttpNamespace

Description

Creates an HTTP namespace. Service instances registered using an HTTP namespace can be discovered using a DiscoverInstances request but can't be discovered using DNS.

For the current quota on the number of namespaces that you can create using the same account, see Cloud Map quotas in the Cloud Map Developer Guide.

Creating a Request

data CreateHttpNamespace

See: newCreateHttpNamespace smart constructor.

Constructors

CreateHttpNamespace'

Fields

creatorRequestId :: Maybe Text

A unique string that identifies the request and that allows failed CreateHttpNamespace requests to be retried without the risk of running the operation twice. CreatorRequestId can be any unique string (for example, a date/time stamp).

tags :: Maybe [Tag]

The tags to add to the namespace. Each tag consists of a key and an optional value that you define. Tags keys can be up to 128 characters in length, and tag values can be up to 256 characters in length.

description :: Maybe Text

A description for the namespace.

name :: Text

The name that you want to assign to this namespace.

newCreateHttpNamespace

Arguments

:: Text
$sel:name:CreateHttpNamespace'

-> CreateHttpNamespace

Create a value of CreateHttpNamespace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creatorRequestId:CreateHttpNamespace', createHttpNamespace_creatorRequestId - A unique string that identifies the request and that allows failed CreateHttpNamespace requests to be retried without the risk of running the operation twice. CreatorRequestId can be any unique string (for example, a date/time stamp).

$sel:tags:CreateHttpNamespace', createHttpNamespace_tags - The tags to add to the namespace. Each tag consists of a key and an optional value that you define. Tags keys can be up to 128 characters in length, and tag values can be up to 256 characters in length.

$sel:description:CreateHttpNamespace', createHttpNamespace_description - A description for the namespace.

$sel:name:CreateHttpNamespace', createHttpNamespace_name - The name that you want to assign to this namespace.

Request Lenses

createHttpNamespace_creatorRequestId :: Lens' CreateHttpNamespace (Maybe Text)

A unique string that identifies the request and that allows failed CreateHttpNamespace requests to be retried without the risk of running the operation twice. CreatorRequestId can be any unique string (for example, a date/time stamp).

createHttpNamespace_tags :: Lens' CreateHttpNamespace (Maybe [Tag])

The tags to add to the namespace. Each tag consists of a key and an optional value that you define. Tags keys can be up to 128 characters in length, and tag values can be up to 256 characters in length.

createHttpNamespace_description :: Lens' CreateHttpNamespace (Maybe Text)

A description for the namespace.

createHttpNamespace_name :: Lens' CreateHttpNamespace Text

The name that you want to assign to this namespace.

Destructuring the Response

data CreateHttpNamespaceResponse

See: newCreateHttpNamespaceResponse smart constructor.

Constructors

CreateHttpNamespaceResponse'

Fields

operationId :: Maybe Text

A value that you can use to determine whether the request completed successfully. To get the status of the operation, see GetOperation.

httpStatus :: Int

The response's http status code.

newCreateHttpNamespaceResponse

Arguments

:: Int
$sel:httpStatus:CreateHttpNamespaceResponse'

-> CreateHttpNamespaceResponse

Create a value of CreateHttpNamespaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationId:CreateHttpNamespaceResponse', createHttpNamespaceResponse_operationId - A value that you can use to determine whether the request completed successfully. To get the status of the operation, see GetOperation.

$sel:httpStatus:CreateHttpNamespaceResponse', createHttpNamespaceResponse_httpStatus - The response's http status code.

Response Lenses

createHttpNamespaceResponse_operationId :: Lens' CreateHttpNamespaceResponse (Maybe Text)

A value that you can use to determine whether the request completed successfully. To get the status of the operation, see GetOperation.

createHttpNamespaceResponse_httpStatus :: Lens' CreateHttpNamespaceResponse Int

The response's http status code.