Network.AWS.Route53AutoNaming.CreatePublicDnsNamespace

Description

Creates a public namespace based on DNS, which is visible on the internet. The namespace defines your service naming scheme. For example, if you name your namespace example.com and name your service backend, the resulting DNS name for the service is backend.example.com. You can discover instances that were registered with a public DNS namespace by using either a DiscoverInstances request or using DNS. For the current quota on the number of namespaces that you can create using the same account, see Cloud Map quotas in the Cloud Map Developer Guide.

Creating a Request

data CreatePublicDnsNamespace

See: newCreatePublicDnsNamespace smart constructor.

Constructors

CreatePublicDnsNamespace'

Fields

creatorRequestId :: Maybe Text

A unique string that identifies the request and that allows failed CreatePublicDnsNamespace requests to be retried without the risk of running the operation twice. CreatorRequestId can be any unique string (for example, a date/timestamp).

description :: Maybe Text

A description for the namespace.

tags :: Maybe [Tag]

The tags to add to the namespace. Each tag consists of a key and an optional value that you define. Tags keys can be up to 128 characters in length, and tag values can be up to 256 characters in length.

properties :: Maybe PublicDnsNamespaceProperties

Properties for the public DNS namespace.

name :: Text

The name that you want to assign to this namespace.

newCreatePublicDnsNamespace

Arguments

:: Text
$sel:name:CreatePublicDnsNamespace'

-> CreatePublicDnsNamespace

Create a value of CreatePublicDnsNamespace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creatorRequestId:CreatePublicDnsNamespace', createPublicDnsNamespace_creatorRequestId - A unique string that identifies the request and that allows failed CreatePublicDnsNamespace requests to be retried without the risk of running the operation twice. CreatorRequestId can be any unique string (for example, a date/timestamp).

$sel:description:CreatePublicDnsNamespace', createPublicDnsNamespace_description - A description for the namespace.

$sel:tags:CreatePublicDnsNamespace', createPublicDnsNamespace_tags - The tags to add to the namespace. Each tag consists of a key and an optional value that you define. Tags keys can be up to 128 characters in length, and tag values can be up to 256 characters in length.

$sel:properties:CreatePublicDnsNamespace', createPublicDnsNamespace_properties - Properties for the public DNS namespace.

$sel:name:CreatePublicDnsNamespace', createPublicDnsNamespace_name - The name that you want to assign to this namespace.

Request Lenses

createPublicDnsNamespace_creatorRequestId :: Lens' CreatePublicDnsNamespace (Maybe Text)

A unique string that identifies the request and that allows failed CreatePublicDnsNamespace requests to be retried without the risk of running the operation twice. CreatorRequestId can be any unique string (for example, a date/timestamp).

createPublicDnsNamespace_description :: Lens' CreatePublicDnsNamespace (Maybe Text)

A description for the namespace.

createPublicDnsNamespace_tags :: Lens' CreatePublicDnsNamespace (Maybe [Tag])

The tags to add to the namespace. Each tag consists of a key and an optional value that you define. Tags keys can be up to 128 characters in length, and tag values can be up to 256 characters in length.

createPublicDnsNamespace_properties :: Lens' CreatePublicDnsNamespace (Maybe PublicDnsNamespaceProperties)

Properties for the public DNS namespace.

createPublicDnsNamespace_name :: Lens' CreatePublicDnsNamespace Text

The name that you want to assign to this namespace.

Destructuring the Response

data CreatePublicDnsNamespaceResponse

See: newCreatePublicDnsNamespaceResponse smart constructor.

Constructors

CreatePublicDnsNamespaceResponse'

Fields

operationId :: Maybe Text

A value that you can use to determine whether the request completed successfully. To get the status of the operation, see GetOperation.

httpStatus :: Int

The response's http status code.

newCreatePublicDnsNamespaceResponse

Arguments

:: Int
$sel:httpStatus:CreatePublicDnsNamespaceResponse'

-> CreatePublicDnsNamespaceResponse

Create a value of CreatePublicDnsNamespaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationId:CreatePublicDnsNamespaceResponse', createPublicDnsNamespaceResponse_operationId - A value that you can use to determine whether the request completed successfully. To get the status of the operation, see GetOperation.

$sel:httpStatus:CreatePublicDnsNamespaceResponse', createPublicDnsNamespaceResponse_httpStatus - The response's http status code.

Response Lenses

createPublicDnsNamespaceResponse_operationId :: Lens' CreatePublicDnsNamespaceResponse (Maybe Text)

A value that you can use to determine whether the request completed successfully. To get the status of the operation, see GetOperation.

createPublicDnsNamespaceResponse_httpStatus :: Lens' CreatePublicDnsNamespaceResponse Int

The response's http status code.