Network.AWS.Route53AutoNaming.DiscoverInstances

Description

Discovers registered instances for a specified namespace and service. You can use DiscoverInstances to discover instances for any type of namespace. For public and private DNS namespaces, you can also use DNS queries to discover instances.

Creating a Request

data DiscoverInstances

See: newDiscoverInstances smart constructor.

Constructors

DiscoverInstances'

Fields

maxResults :: Maybe Natural

The maximum number of instances that you want Cloud Map to return in the response to a DiscoverInstances request. If you don't specify a value for MaxResults, Cloud Map returns up to 100 instances.

optionalParameters :: Maybe (HashMap Text Text)

Opportunistic filters to scope the results based on custom attributes. If there are instances that match both the filters specified in both the QueryParameters parameter and this parameter, all of these instances are returned. Otherwise, the filters are ignored, and only instances that match the filters that are specified in the QueryParameters parameter are returned.

queryParameters :: Maybe (HashMap Text Text)

Filters to scope the results based on custom attributes for the instance (for example, {version=v1, az=1a}). Only instances that match all the specified key-value pairs are returned.

healthStatus :: Maybe HealthStatusFilter

The health status of the instances that you want to discover. This parameter is ignored for services that don't have a health check configured, and all instances are returned.

HEALTHY

Returns healthy instances.

UNHEALTHY

Returns unhealthy instances.

ALL

Returns all instances.

HEALTHY_OR_ELSE_ALL

Returns healthy instances, unless none are reporting a healthy state. In that case, return all instances. This is also called failing open.

namespaceName :: Text

The HttpName name of the namespace. It's found in the HttpProperties member of the Properties member of the namespace.

serviceName :: Text

The name of the service that you specified when you registered the instance.

newDiscoverInstances

Arguments

:: Text
$sel:namespaceName:DiscoverInstances'

-> Text
$sel:serviceName:DiscoverInstances'

-> DiscoverInstances

Create a value of DiscoverInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxResults:DiscoverInstances', discoverInstances_maxResults - The maximum number of instances that you want Cloud Map to return in the response to a DiscoverInstances request. If you don't specify a value for MaxResults, Cloud Map returns up to 100 instances.

$sel:optionalParameters:DiscoverInstances', discoverInstances_optionalParameters - Opportunistic filters to scope the results based on custom attributes. If there are instances that match both the filters specified in both the QueryParameters parameter and this parameter, all of these instances are returned. Otherwise, the filters are ignored, and only instances that match the filters that are specified in the QueryParameters parameter are returned.

$sel:queryParameters:DiscoverInstances', discoverInstances_queryParameters - Filters to scope the results based on custom attributes for the instance (for example, {version=v1, az=1a}). Only instances that match all the specified key-value pairs are returned.

$sel:healthStatus:DiscoverInstances', discoverInstances_healthStatus - The health status of the instances that you want to discover. This parameter is ignored for services that don't have a health check configured, and all instances are returned.

HEALTHY

Returns healthy instances.

UNHEALTHY

Returns unhealthy instances.

ALL

Returns all instances.

HEALTHY_OR_ELSE_ALL

Returns healthy instances, unless none are reporting a healthy state. In that case, return all instances. This is also called failing open.

$sel:namespaceName:DiscoverInstances', discoverInstances_namespaceName - The HttpName name of the namespace. It's found in the HttpProperties member of the Properties member of the namespace.

$sel:serviceName:DiscoverInstances', discoverInstances_serviceName - The name of the service that you specified when you registered the instance.

Request Lenses

discoverInstances_maxResults :: Lens' DiscoverInstances (Maybe Natural)

The maximum number of instances that you want Cloud Map to return in the response to a DiscoverInstances request. If you don't specify a value for MaxResults, Cloud Map returns up to 100 instances.

discoverInstances_optionalParameters :: Lens' DiscoverInstances (Maybe (HashMap Text Text))

Opportunistic filters to scope the results based on custom attributes. If there are instances that match both the filters specified in both the QueryParameters parameter and this parameter, all of these instances are returned. Otherwise, the filters are ignored, and only instances that match the filters that are specified in the QueryParameters parameter are returned.

discoverInstances_queryParameters :: Lens' DiscoverInstances (Maybe (HashMap Text Text))

Filters to scope the results based on custom attributes for the instance (for example, {version=v1, az=1a}). Only instances that match all the specified key-value pairs are returned.

discoverInstances_healthStatus :: Lens' DiscoverInstances (Maybe HealthStatusFilter)

The health status of the instances that you want to discover. This parameter is ignored for services that don't have a health check configured, and all instances are returned.

HEALTHY

Returns healthy instances.

UNHEALTHY

Returns unhealthy instances.

ALL

Returns all instances.

HEALTHY_OR_ELSE_ALL

Returns healthy instances, unless none are reporting a healthy state. In that case, return all instances. This is also called failing open.

discoverInstances_namespaceName :: Lens' DiscoverInstances Text

The HttpName name of the namespace. It's found in the HttpProperties member of the Properties member of the namespace.

discoverInstances_serviceName :: Lens' DiscoverInstances Text

The name of the service that you specified when you registered the instance.

Destructuring the Response

data DiscoverInstancesResponse

See: newDiscoverInstancesResponse smart constructor.

Constructors

DiscoverInstancesResponse'

Fields

instances :: Maybe [HttpInstanceSummary]

A complex type that contains one HttpInstanceSummary for each registered instance.

httpStatus :: Int

The response's http status code.

newDiscoverInstancesResponse

Arguments

:: Int
$sel:httpStatus:DiscoverInstancesResponse'

-> DiscoverInstancesResponse

Create a value of DiscoverInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instances:DiscoverInstancesResponse', discoverInstancesResponse_instances - A complex type that contains one HttpInstanceSummary for each registered instance.

$sel:httpStatus:DiscoverInstancesResponse', discoverInstancesResponse_httpStatus - The response's http status code.

Response Lenses

discoverInstancesResponse_instances :: Lens' DiscoverInstancesResponse (Maybe [HttpInstanceSummary])

A complex type that contains one HttpInstanceSummary for each registered instance.

discoverInstancesResponse_httpStatus :: Lens' DiscoverInstancesResponse Int

The response's http status code.