Network.AWS.Route53AutoNaming.Types.HttpInstanceSummary

Description

Documentation

data HttpInstanceSummary

In a response to a DiscoverInstances request, HttpInstanceSummary contains information about one instance that matches the values that you specified in the request.

See: newHttpInstanceSummary smart constructor.

Constructors

HttpInstanceSummary'

Fields

instanceId :: Maybe Text

The ID of an instance that matches the values that you specified in the request.

namespaceName :: Maybe Text

@ @

The HttpName name of the namespace. It's found in the HttpProperties member of the Properties member of the namespace.

serviceName :: Maybe Text

The name of the service that you specified when you registered the instance.

attributes :: Maybe (HashMap Text Text)

If you included any attributes when you registered the instance, the values of those attributes.

healthStatus :: Maybe HealthStatus

If you configured health checking in the service, the current health status of the service instance.

newHttpInstanceSummary :: HttpInstanceSummary

Create a value of HttpInstanceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:HttpInstanceSummary', httpInstanceSummary_instanceId - The ID of an instance that matches the values that you specified in the request.

$sel:namespaceName:HttpInstanceSummary', httpInstanceSummary_namespaceName - @ @

The HttpName name of the namespace. It's found in the HttpProperties member of the Properties member of the namespace.

$sel:serviceName:HttpInstanceSummary', httpInstanceSummary_serviceName - The name of the service that you specified when you registered the instance.

$sel:attributes:HttpInstanceSummary', httpInstanceSummary_attributes - If you included any attributes when you registered the instance, the values of those attributes.

$sel:healthStatus:HttpInstanceSummary', httpInstanceSummary_healthStatus - If you configured health checking in the service, the current health status of the service instance.

httpInstanceSummary_instanceId :: Lens' HttpInstanceSummary (Maybe Text)

The ID of an instance that matches the values that you specified in the request.

httpInstanceSummary_namespaceName :: Lens' HttpInstanceSummary (Maybe Text)

@ @

The HttpName name of the namespace. It's found in the HttpProperties member of the Properties member of the namespace.

httpInstanceSummary_serviceName :: Lens' HttpInstanceSummary (Maybe Text)

The name of the service that you specified when you registered the instance.

httpInstanceSummary_attributes :: Lens' HttpInstanceSummary (Maybe (HashMap Text Text))

If you included any attributes when you registered the instance, the values of those attributes.

httpInstanceSummary_healthStatus :: Lens' HttpInstanceSummary (Maybe HealthStatus)

If you configured health checking in the service, the current health status of the service instance.