Network.AWS.Route53AutoNaming.Types.Namespace

Description

Documentation

data Namespace

A complex type that contains information about a specified namespace.

See: newNamespace smart constructor.

Constructors

Namespace'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) that Cloud Map assigns to the namespace when you create it.

creatorRequestId :: Maybe Text

A unique string that identifies the request and that allows failed requests to be retried without the risk of running an operation twice.

createDate :: Maybe POSIX

The date that the namespace was created, in Unix date/time format and Coordinated Universal Time (UTC). The value of CreateDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

serviceCount :: Maybe Int

The number of services that are associated with the namespace.

name :: Maybe Text

The name of the namespace, such as example.com.

id :: Maybe Text

The ID of a namespace.

type' :: Maybe NamespaceType

The type of the namespace. The methods for discovering instances depends on the value that you specify:

HTTP

Instances can be discovered only programmatically, using the Cloud Map DiscoverInstances API.

DNS_PUBLIC

Instances can be discovered using public DNS queries and using the DiscoverInstances API.

DNS_PRIVATE

Instances can be discovered using DNS queries in VPCs and using the DiscoverInstances API.

description :: Maybe Text

The description that you specify for the namespace when you create it.

properties :: Maybe NamespaceProperties

A complex type that contains information that's specific to the type of the namespace.

newNamespace :: Namespace

Create a value of Namespace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Namespace', namespace_arn - The Amazon Resource Name (ARN) that Cloud Map assigns to the namespace when you create it.

$sel:creatorRequestId:Namespace', namespace_creatorRequestId - A unique string that identifies the request and that allows failed requests to be retried without the risk of running an operation twice.

$sel:createDate:Namespace', namespace_createDate - The date that the namespace was created, in Unix date/time format and Coordinated Universal Time (UTC). The value of CreateDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:serviceCount:Namespace', namespace_serviceCount - The number of services that are associated with the namespace.

$sel:name:Namespace', namespace_name - The name of the namespace, such as example.com.

$sel:id:Namespace', namespace_id - The ID of a namespace.

$sel:type':Namespace', namespace_type - The type of the namespace. The methods for discovering instances depends on the value that you specify:

HTTP

Instances can be discovered only programmatically, using the Cloud Map DiscoverInstances API.

DNS_PUBLIC

Instances can be discovered using public DNS queries and using the DiscoverInstances API.

DNS_PRIVATE

Instances can be discovered using DNS queries in VPCs and using the DiscoverInstances API.

$sel:description:Namespace', namespace_description - The description that you specify for the namespace when you create it.

$sel:properties:Namespace', namespace_properties - A complex type that contains information that's specific to the type of the namespace.

namespace_arn :: Lens' Namespace (Maybe Text)

The Amazon Resource Name (ARN) that Cloud Map assigns to the namespace when you create it.

namespace_creatorRequestId :: Lens' Namespace (Maybe Text)

A unique string that identifies the request and that allows failed requests to be retried without the risk of running an operation twice.

namespace_createDate :: Lens' Namespace (Maybe UTCTime)

The date that the namespace was created, in Unix date/time format and Coordinated Universal Time (UTC). The value of CreateDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

namespace_serviceCount :: Lens' Namespace (Maybe Int)

The number of services that are associated with the namespace.

namespace_name :: Lens' Namespace (Maybe Text)

The name of the namespace, such as example.com.

namespace_id :: Lens' Namespace (Maybe Text)

The ID of a namespace.

namespace_type :: Lens' Namespace (Maybe NamespaceType)

The type of the namespace. The methods for discovering instances depends on the value that you specify:

HTTP

Instances can be discovered only programmatically, using the Cloud Map DiscoverInstances API.

DNS_PUBLIC

Instances can be discovered using public DNS queries and using the DiscoverInstances API.

DNS_PRIVATE

Instances can be discovered using DNS queries in VPCs and using the DiscoverInstances API.

namespace_description :: Lens' Namespace (Maybe Text)

The description that you specify for the namespace when you create it.

namespace_properties :: Lens' Namespace (Maybe NamespaceProperties)

A complex type that contains information that's specific to the type of the namespace.