Network.AWS.Route53AutoNaming.Types.NamespaceSummary

Description

Documentation

data NamespaceSummary

A complex type that contains information about a namespace.

See: newNamespaceSummary smart constructor.

Constructors

NamespaceSummary'

Fields

createDate :: Maybe POSIX

The date and time that the namespace was created.

arn :: Maybe Text

The Amazon Resource Name (ARN) that Cloud Map assigns to the namespace when you create it.

id :: Maybe Text

The ID of the namespace.

name :: Maybe Text

The name of the namespace. When you create a namespace, Cloud Map automatically creates a Route 53 hosted zone that has the same name as the namespace.

properties :: Maybe NamespaceProperties

The properties of the namespace.

description :: Maybe Text

A description for the namespace.

serviceCount :: Maybe Int

The number of services that were created using the namespace.

type' :: Maybe NamespaceType

The type of the namespace, either public or private.

newNamespaceSummary :: NamespaceSummary

Create a value of NamespaceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createDate:NamespaceSummary', namespaceSummary_createDate - The date and time that the namespace was created.

$sel:arn:NamespaceSummary', namespaceSummary_arn - The Amazon Resource Name (ARN) that Cloud Map assigns to the namespace when you create it.

$sel:id:NamespaceSummary', namespaceSummary_id - The ID of the namespace.

$sel:name:NamespaceSummary', namespaceSummary_name - The name of the namespace. When you create a namespace, Cloud Map automatically creates a Route 53 hosted zone that has the same name as the namespace.

$sel:properties:NamespaceSummary', namespaceSummary_properties - The properties of the namespace.

$sel:description:NamespaceSummary', namespaceSummary_description - A description for the namespace.

$sel:serviceCount:NamespaceSummary', namespaceSummary_serviceCount - The number of services that were created using the namespace.

$sel:type':NamespaceSummary', namespaceSummary_type - The type of the namespace, either public or private.

namespaceSummary_createDate :: Lens' NamespaceSummary (Maybe UTCTime)

The date and time that the namespace was created.

namespaceSummary_arn :: Lens' NamespaceSummary (Maybe Text)

The Amazon Resource Name (ARN) that Cloud Map assigns to the namespace when you create it.

namespaceSummary_id :: Lens' NamespaceSummary (Maybe Text)

The ID of the namespace.

namespaceSummary_name :: Lens' NamespaceSummary (Maybe Text)

The name of the namespace. When you create a namespace, Cloud Map automatically creates a Route 53 hosted zone that has the same name as the namespace.

namespaceSummary_properties :: Lens' NamespaceSummary (Maybe NamespaceProperties)

The properties of the namespace.

namespaceSummary_description :: Lens' NamespaceSummary (Maybe Text)

A description for the namespace.

namespaceSummary_serviceCount :: Lens' NamespaceSummary (Maybe Int)

The number of services that were created using the namespace.

namespaceSummary_type :: Lens' NamespaceSummary (Maybe NamespaceType)

The type of the namespace, either public or private.