Network.AWS.Route53AutoNaming.Types.ServiceInfo

Description

Documentation

data ServiceInfo

A complex type that contains information about the specified service.

See: newServiceInfo smart constructor.

Constructors

ServiceInfo'

Fields

instanceCount :: Maybe Int

The number of instances that are currently associated with the service. Instances that were previously associated with the service but that are deleted aren't included in the count. The count might not reflect pending registrations and deregistrations.

arn :: Maybe Text

The Amazon Resource Name (ARN) that Cloud Map assigns to the service when you create it.

healthCheckConfig :: Maybe HealthCheckConfig

Public DNS and HTTP namespaces only. A complex type that contains settings for an optional health check. If you specify settings for a health check, Cloud Map associates the health check with the records that you specify in DnsConfig.

For information about the charges for health checks, see Amazon Route 53 Pricing.

creatorRequestId :: Maybe Text

A unique string that identifies the request and that allows failed requests to be retried without the risk of running the operation twice. CreatorRequestId can be any unique string (for example, a date/timestamp).

createDate :: Maybe POSIX

The date and time that the service was created, in Unix format and Coordinated Universal Time (UTC). The value of CreateDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

healthCheckCustomConfig :: Maybe HealthCheckCustomConfig

A complex type that contains information about an optional custom health check.

If you specify a health check configuration, you can specify either HealthCheckCustomConfig or HealthCheckConfig but not both.

namespaceId :: Maybe Text

The ID of the namespace that was used to create the service.

name :: Maybe Text

The name of the service.

id :: Maybe Text

The ID that Cloud Map assigned to the service when you created it.

type' :: Maybe ServiceType

Describes the systems that can be used to discover the service instances.

DNS_HTTP

The service instances can be discovered using either DNS queries or the DiscoverInstances API operation.

HTTP

The service instances can only be discovered using the DiscoverInstances API operation.

DNS

Reserved.

dnsConfig :: Maybe DnsConfig

A complex type that contains information about the Route 53 DNS records that you want Cloud Map to create when you register an instance.

description :: Maybe Text

The description of the service.

newServiceInfo :: ServiceInfo

Create a value of ServiceInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceCount:ServiceInfo', serviceInfo_instanceCount - The number of instances that are currently associated with the service. Instances that were previously associated with the service but that are deleted aren't included in the count. The count might not reflect pending registrations and deregistrations.

$sel:arn:ServiceInfo', serviceInfo_arn - The Amazon Resource Name (ARN) that Cloud Map assigns to the service when you create it.

$sel:healthCheckConfig:ServiceInfo', serviceInfo_healthCheckConfig - Public DNS and HTTP namespaces only. A complex type that contains settings for an optional health check. If you specify settings for a health check, Cloud Map associates the health check with the records that you specify in DnsConfig.

For information about the charges for health checks, see Amazon Route 53 Pricing.

$sel:creatorRequestId:ServiceInfo', serviceInfo_creatorRequestId - A unique string that identifies the request and that allows failed requests to be retried without the risk of running the operation twice. CreatorRequestId can be any unique string (for example, a date/timestamp).

$sel:createDate:ServiceInfo', serviceInfo_createDate - The date and time that the service was created, in Unix format and Coordinated Universal Time (UTC). The value of CreateDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:healthCheckCustomConfig:ServiceInfo', serviceInfo_healthCheckCustomConfig - A complex type that contains information about an optional custom health check.

If you specify a health check configuration, you can specify either HealthCheckCustomConfig or HealthCheckConfig but not both.

$sel:namespaceId:ServiceInfo', serviceInfo_namespaceId - The ID of the namespace that was used to create the service.

$sel:name:ServiceInfo', serviceInfo_name - The name of the service.

$sel:id:ServiceInfo', serviceInfo_id - The ID that Cloud Map assigned to the service when you created it.

$sel:type':ServiceInfo', serviceInfo_type - Describes the systems that can be used to discover the service instances.

DNS_HTTP

The service instances can be discovered using either DNS queries or the DiscoverInstances API operation.

HTTP

The service instances can only be discovered using the DiscoverInstances API operation.

DNS

Reserved.

$sel:dnsConfig:ServiceInfo', serviceInfo_dnsConfig - A complex type that contains information about the Route 53 DNS records that you want Cloud Map to create when you register an instance.

$sel:description:ServiceInfo', serviceInfo_description - The description of the service.

serviceInfo_instanceCount :: Lens' ServiceInfo (Maybe Int)

The number of instances that are currently associated with the service. Instances that were previously associated with the service but that are deleted aren't included in the count. The count might not reflect pending registrations and deregistrations.

serviceInfo_arn :: Lens' ServiceInfo (Maybe Text)

The Amazon Resource Name (ARN) that Cloud Map assigns to the service when you create it.

serviceInfo_healthCheckConfig :: Lens' ServiceInfo (Maybe HealthCheckConfig)

Public DNS and HTTP namespaces only. A complex type that contains settings for an optional health check. If you specify settings for a health check, Cloud Map associates the health check with the records that you specify in DnsConfig.

For information about the charges for health checks, see Amazon Route 53 Pricing.

serviceInfo_creatorRequestId :: Lens' ServiceInfo (Maybe Text)

A unique string that identifies the request and that allows failed requests to be retried without the risk of running the operation twice. CreatorRequestId can be any unique string (for example, a date/timestamp).

serviceInfo_createDate :: Lens' ServiceInfo (Maybe UTCTime)

The date and time that the service was created, in Unix format and Coordinated Universal Time (UTC). The value of CreateDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

serviceInfo_healthCheckCustomConfig :: Lens' ServiceInfo (Maybe HealthCheckCustomConfig)

A complex type that contains information about an optional custom health check.

If you specify a health check configuration, you can specify either HealthCheckCustomConfig or HealthCheckConfig but not both.

serviceInfo_namespaceId :: Lens' ServiceInfo (Maybe Text)

The ID of the namespace that was used to create the service.

serviceInfo_name :: Lens' ServiceInfo (Maybe Text)

The name of the service.

serviceInfo_id :: Lens' ServiceInfo (Maybe Text)

The ID that Cloud Map assigned to the service when you created it.

serviceInfo_type :: Lens' ServiceInfo (Maybe ServiceType)

Describes the systems that can be used to discover the service instances.

DNS_HTTP

The service instances can be discovered using either DNS queries or the DiscoverInstances API operation.

HTTP

The service instances can only be discovered using the DiscoverInstances API operation.

DNS

Reserved.

serviceInfo_dnsConfig :: Lens' ServiceInfo (Maybe DnsConfig)

A complex type that contains information about the Route 53 DNS records that you want Cloud Map to create when you register an instance.

serviceInfo_description :: Lens' ServiceInfo (Maybe Text)

The description of the service.