Network.AWS.Route53AutoNaming.Types.ServiceSummary

Description

Documentation

data ServiceSummary

A complex type that contains information about a specified service.

See: newServiceSummary smart constructor.

Constructors

ServiceSummary'

Fields

dnsConfig :: Maybe DnsConfig

Information about the Route 53 DNS records that you want Cloud Map to create when you register an instance.

createDate :: Maybe POSIX

The date and time that the service was created.

arn :: Maybe Text

The Amazon Resource Name (ARN) that Cloud Map assigns to the service when you create it.

id :: Maybe Text

The ID that Cloud Map assigned to the service when you created it.

name :: Maybe Text

The name of the service.

description :: Maybe Text

The description that you specify when you create the service.

healthCheckCustomConfig :: Maybe HealthCheckCustomConfig

Information about an optional custom health check. A custom health check, which requires that you use a third-party health checker to evaluate the health of your resources, is useful in the following circumstances:

You can't use a health check that's defined by HealthCheckConfig because the resource isn't available over the internet. For example, you can use a custom health check when the instance is in an Amazon VPC. (To check the health of resources in a VPC, the health checker must also be in the VPC.)

You want to use a third-party health checker regardless of where your resources are located.

If you specify a health check configuration, you can specify either HealthCheckCustomConfig or HealthCheckConfig but not both.

healthCheckConfig :: Maybe HealthCheckConfig

Public DNS and HTTP namespaces only. Settings for an optional health check. If you specify settings for a health check, Cloud Map associates the health check with the records that you specify in DnsConfig.

type' :: Maybe ServiceType

Describes the systems that can be used to discover the service instances.

DNS_HTTP

The service instances can be discovered using either DNS queries or the DiscoverInstances API operation.

HTTP

The service instances can only be discovered using the DiscoverInstances API operation.

DNS

Reserved.

instanceCount :: Maybe Int

The number of instances that are currently associated with the service. Instances that were previously associated with the service but that are deleted aren't included in the count. The count might not reflect pending registrations and deregistrations.

newServiceSummary :: ServiceSummary

Create a value of ServiceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dnsConfig:ServiceSummary', serviceSummary_dnsConfig - Information about the Route 53 DNS records that you want Cloud Map to create when you register an instance.

$sel:createDate:ServiceSummary', serviceSummary_createDate - The date and time that the service was created.

$sel:arn:ServiceSummary', serviceSummary_arn - The Amazon Resource Name (ARN) that Cloud Map assigns to the service when you create it.

$sel:id:ServiceSummary', serviceSummary_id - The ID that Cloud Map assigned to the service when you created it.

$sel:name:ServiceSummary', serviceSummary_name - The name of the service.

$sel:description:ServiceSummary', serviceSummary_description - The description that you specify when you create the service.

$sel:healthCheckCustomConfig:ServiceSummary', serviceSummary_healthCheckCustomConfig - Information about an optional custom health check. A custom health check, which requires that you use a third-party health checker to evaluate the health of your resources, is useful in the following circumstances:

You can't use a health check that's defined by HealthCheckConfig because the resource isn't available over the internet. For example, you can use a custom health check when the instance is in an Amazon VPC. (To check the health of resources in a VPC, the health checker must also be in the VPC.)

You want to use a third-party health checker regardless of where your resources are located.

If you specify a health check configuration, you can specify either HealthCheckCustomConfig or HealthCheckConfig but not both.

$sel:healthCheckConfig:ServiceSummary', serviceSummary_healthCheckConfig - Public DNS and HTTP namespaces only. Settings for an optional health check. If you specify settings for a health check, Cloud Map associates the health check with the records that you specify in DnsConfig.

$sel:type':ServiceSummary', serviceSummary_type - Describes the systems that can be used to discover the service instances.

DNS_HTTP

The service instances can be discovered using either DNS queries or the DiscoverInstances API operation.

HTTP

The service instances can only be discovered using the DiscoverInstances API operation.

DNS

Reserved.

$sel:instanceCount:ServiceSummary', serviceSummary_instanceCount - The number of instances that are currently associated with the service. Instances that were previously associated with the service but that are deleted aren't included in the count. The count might not reflect pending registrations and deregistrations.

serviceSummary_dnsConfig :: Lens' ServiceSummary (Maybe DnsConfig)

Information about the Route 53 DNS records that you want Cloud Map to create when you register an instance.

serviceSummary_createDate :: Lens' ServiceSummary (Maybe UTCTime)

The date and time that the service was created.

serviceSummary_arn :: Lens' ServiceSummary (Maybe Text)

The Amazon Resource Name (ARN) that Cloud Map assigns to the service when you create it.

serviceSummary_id :: Lens' ServiceSummary (Maybe Text)

The ID that Cloud Map assigned to the service when you created it.

serviceSummary_name :: Lens' ServiceSummary (Maybe Text)

The name of the service.

serviceSummary_description :: Lens' ServiceSummary (Maybe Text)

The description that you specify when you create the service.

serviceSummary_healthCheckCustomConfig :: Lens' ServiceSummary (Maybe HealthCheckCustomConfig)

Information about an optional custom health check. A custom health check, which requires that you use a third-party health checker to evaluate the health of your resources, is useful in the following circumstances:

You can't use a health check that's defined by HealthCheckConfig because the resource isn't available over the internet. For example, you can use a custom health check when the instance is in an Amazon VPC. (To check the health of resources in a VPC, the health checker must also be in the VPC.)

You want to use a third-party health checker regardless of where your resources are located.

If you specify a health check configuration, you can specify either HealthCheckCustomConfig or HealthCheckConfig but not both.

serviceSummary_healthCheckConfig :: Lens' ServiceSummary (Maybe HealthCheckConfig)

Public DNS and HTTP namespaces only. Settings for an optional health check. If you specify settings for a health check, Cloud Map associates the health check with the records that you specify in DnsConfig.

serviceSummary_type :: Lens' ServiceSummary (Maybe ServiceType)

Describes the systems that can be used to discover the service instances.

DNS_HTTP

The service instances can be discovered using either DNS queries or the DiscoverInstances API operation.

HTTP

The service instances can only be discovered using the DiscoverInstances API operation.

DNS

Reserved.

serviceSummary_instanceCount :: Lens' ServiceSummary (Maybe Int)

The number of instances that are currently associated with the service. Instances that were previously associated with the service but that are deleted aren't included in the count. The count might not reflect pending registrations and deregistrations.