Network.AWS.Route53AutoNaming.UpdateInstanceCustomHealthStatus

Description

Submits a request to change the health status of a custom health check to healthy or unhealthy.

You can use UpdateInstanceCustomHealthStatus to change the status only for custom health checks, which you define using HealthCheckCustomConfig when you create a service. You can't use it to change the status for Route 53 health checks, which you define using HealthCheckConfig.

For more information, see HealthCheckCustomConfig.

Creating a Request

data UpdateInstanceCustomHealthStatus

See: newUpdateInstanceCustomHealthStatus smart constructor.

Constructors

UpdateInstanceCustomHealthStatus'

Fields

serviceId :: Text

The ID of the service that includes the configuration for the custom health check that you want to change the status for.

instanceId :: Text

The ID of the instance that you want to change the health status for.

status :: CustomHealthStatus

The new status of the instance, HEALTHY or UNHEALTHY.

newUpdateInstanceCustomHealthStatus

Arguments

:: Text
$sel:serviceId:UpdateInstanceCustomHealthStatus'

-> Text
$sel:instanceId:UpdateInstanceCustomHealthStatus'

-> CustomHealthStatus
$sel:status:UpdateInstanceCustomHealthStatus'

-> UpdateInstanceCustomHealthStatus

Create a value of UpdateInstanceCustomHealthStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceId:UpdateInstanceCustomHealthStatus', updateInstanceCustomHealthStatus_serviceId - The ID of the service that includes the configuration for the custom health check that you want to change the status for.

$sel:instanceId:UpdateInstanceCustomHealthStatus', updateInstanceCustomHealthStatus_instanceId - The ID of the instance that you want to change the health status for.

$sel:status:UpdateInstanceCustomHealthStatus', updateInstanceCustomHealthStatus_status - The new status of the instance, HEALTHY or UNHEALTHY.

Request Lenses

updateInstanceCustomHealthStatus_serviceId :: Lens' UpdateInstanceCustomHealthStatus Text

The ID of the service that includes the configuration for the custom health check that you want to change the status for.

updateInstanceCustomHealthStatus_instanceId :: Lens' UpdateInstanceCustomHealthStatus Text

The ID of the instance that you want to change the health status for.

updateInstanceCustomHealthStatus_status :: Lens' UpdateInstanceCustomHealthStatus CustomHealthStatus

The new status of the instance, HEALTHY or UNHEALTHY.

Destructuring the Response

data UpdateInstanceCustomHealthStatusResponse

See: newUpdateInstanceCustomHealthStatusResponse smart constructor.

Constructors

UpdateInstanceCustomHealthStatusResponse'

newUpdateInstanceCustomHealthStatusResponse :: UpdateInstanceCustomHealthStatusResponse

Create a value of UpdateInstanceCustomHealthStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.