Network.AWS.Route53AutoNaming.UpdatePublicDnsNamespace

Description

Updates a public DNS namespace.

Creating a Request

data UpdatePublicDnsNamespace

See: newUpdatePublicDnsNamespace smart constructor.

Constructors

UpdatePublicDnsNamespace'

Fields

updaterRequestId :: Maybe Text

A unique string that identifies the request and that allows failed UpdatePublicDnsNamespace requests to be retried without the risk of running the operation twice. UpdaterRequestId can be any unique string (for example, a date/timestamp).

id :: Text

The ID of the namespace being updated.

namespace :: PublicDnsNamespaceChange

Updated properties for the public DNS namespace.

newUpdatePublicDnsNamespace

Arguments

:: Text
$sel:id:UpdatePublicDnsNamespace'

-> PublicDnsNamespaceChange
$sel:namespace:UpdatePublicDnsNamespace'

-> UpdatePublicDnsNamespace

Create a value of UpdatePublicDnsNamespace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updaterRequestId:UpdatePublicDnsNamespace', updatePublicDnsNamespace_updaterRequestId - A unique string that identifies the request and that allows failed UpdatePublicDnsNamespace requests to be retried without the risk of running the operation twice. UpdaterRequestId can be any unique string (for example, a date/timestamp).

$sel:id:UpdatePublicDnsNamespace', updatePublicDnsNamespace_id - The ID of the namespace being updated.

$sel:namespace:UpdatePublicDnsNamespace', updatePublicDnsNamespace_namespace - Updated properties for the public DNS namespace.

Request Lenses

updatePublicDnsNamespace_updaterRequestId :: Lens' UpdatePublicDnsNamespace (Maybe Text)

A unique string that identifies the request and that allows failed UpdatePublicDnsNamespace requests to be retried without the risk of running the operation twice. UpdaterRequestId can be any unique string (for example, a date/timestamp).

updatePublicDnsNamespace_id :: Lens' UpdatePublicDnsNamespace Text

The ID of the namespace being updated.

updatePublicDnsNamespace_namespace :: Lens' UpdatePublicDnsNamespace PublicDnsNamespaceChange

Updated properties for the public DNS namespace.

Destructuring the Response

data UpdatePublicDnsNamespaceResponse

See: newUpdatePublicDnsNamespaceResponse smart constructor.

Constructors

UpdatePublicDnsNamespaceResponse'

Fields

operationId :: Maybe Text

A value that you can use to determine whether the request completed successfully. To get the status of the operation, see GetOperation.

httpStatus :: Int

The response's http status code.

newUpdatePublicDnsNamespaceResponse

Arguments

:: Int
$sel:httpStatus:UpdatePublicDnsNamespaceResponse'

-> UpdatePublicDnsNamespaceResponse

Create a value of UpdatePublicDnsNamespaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationId:UpdatePublicDnsNamespaceResponse', updatePublicDnsNamespaceResponse_operationId - A value that you can use to determine whether the request completed successfully. To get the status of the operation, see GetOperation.

$sel:httpStatus:UpdatePublicDnsNamespaceResponse', updatePublicDnsNamespaceResponse_httpStatus - The response's http status code.

Response Lenses

updatePublicDnsNamespaceResponse_operationId :: Lens' UpdatePublicDnsNamespaceResponse (Maybe Text)

A value that you can use to determine whether the request completed successfully. To get the status of the operation, see GetOperation.

updatePublicDnsNamespaceResponse_httpStatus :: Lens' UpdatePublicDnsNamespaceResponse Int

The response's http status code.