Network.AWS.Route53Domains.CancelDomainTransferToAnotherAwsAccount

Description

Cancels the transfer of a domain from the current AWS account to another AWS account. You initiate a transfer between AWS accounts using TransferDomainToAnotherAwsAccount.

You must cancel the transfer before the other AWS account accepts the transfer using AcceptDomainTransferFromAnotherAwsAccount.

Use either ListOperations or GetOperationDetail to determine whether the operation succeeded. GetOperationDetail provides additional information, for example, Domain Transfer from Aws Account 111122223333 has been cancelled.

Creating a Request

data CancelDomainTransferToAnotherAwsAccount

The CancelDomainTransferToAnotherAwsAccount request includes the following element.

See: newCancelDomainTransferToAnotherAwsAccount smart constructor.

Constructors

CancelDomainTransferToAnotherAwsAccount'

Fields

domainName :: Text

The name of the domain for which you want to cancel the transfer to another AWS account.

newCancelDomainTransferToAnotherAwsAccount

Arguments

:: Text
$sel:domainName:CancelDomainTransferToAnotherAwsAccount'

-> CancelDomainTransferToAnotherAwsAccount

Create a value of CancelDomainTransferToAnotherAwsAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:CancelDomainTransferToAnotherAwsAccount', cancelDomainTransferToAnotherAwsAccount_domainName - The name of the domain for which you want to cancel the transfer to another AWS account.

Request Lenses

cancelDomainTransferToAnotherAwsAccount_domainName :: Lens' CancelDomainTransferToAnotherAwsAccount Text

The name of the domain for which you want to cancel the transfer to another AWS account.

Destructuring the Response

data CancelDomainTransferToAnotherAwsAccountResponse

The CancelDomainTransferToAnotherAwsAccount response includes the following element.

See: newCancelDomainTransferToAnotherAwsAccountResponse smart constructor.

Constructors

CancelDomainTransferToAnotherAwsAccountResponse'

Fields

operationId :: Maybe Text

The identifier that TransferDomainToAnotherAwsAccount returned to track the progress of the request. Because the transfer request was canceled, the value is no longer valid, and you can't use GetOperationDetail to query the operation status.

httpStatus :: Int

The response's http status code.

newCancelDomainTransferToAnotherAwsAccountResponse

Arguments

:: Int
$sel:httpStatus:CancelDomainTransferToAnotherAwsAccountResponse'

-> CancelDomainTransferToAnotherAwsAccountResponse

Create a value of CancelDomainTransferToAnotherAwsAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationId:CancelDomainTransferToAnotherAwsAccountResponse', cancelDomainTransferToAnotherAwsAccountResponse_operationId - The identifier that TransferDomainToAnotherAwsAccount returned to track the progress of the request. Because the transfer request was canceled, the value is no longer valid, and you can't use GetOperationDetail to query the operation status.

$sel:httpStatus:CancelDomainTransferToAnotherAwsAccountResponse', cancelDomainTransferToAnotherAwsAccountResponse_httpStatus - The response's http status code.

Response Lenses

cancelDomainTransferToAnotherAwsAccountResponse_operationId :: Lens' CancelDomainTransferToAnotherAwsAccountResponse (Maybe Text)

The identifier that TransferDomainToAnotherAwsAccount returned to track the progress of the request. Because the transfer request was canceled, the value is no longer valid, and you can't use GetOperationDetail to query the operation status.

cancelDomainTransferToAnotherAwsAccountResponse_httpStatus :: Lens' CancelDomainTransferToAnotherAwsAccountResponse Int

The response's http status code.