Network.AWS.Route53Domains.CheckDomainTransferability

Description

Checks whether a domain name can be transferred to Amazon Route 53.

Creating a Request

data CheckDomainTransferability

The CheckDomainTransferability request contains the following elements.

See: newCheckDomainTransferability smart constructor.

Constructors

CheckDomainTransferability'

Fields

authCode :: Maybe (Sensitive Text)

If the registrar for the top-level domain (TLD) requires an authorization code to transfer the domain, the code that you got from the current registrar for the domain.

domainName :: Text

The name of the domain that you want to transfer to Route 53. The top-level domain (TLD), such as .com, must be a TLD that Route 53 supports. For a list of supported TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

The domain name can contain only the following characters:

Letters a through z. Domain names are not case sensitive.

Numbers 0 through 9.

Hyphen (-). You can't specify a hyphen at the beginning or end of a label.

Period (.) to separate the labels in the name, such as the . in example.com.

newCheckDomainTransferability

Arguments

:: Text
$sel:domainName:CheckDomainTransferability'

-> CheckDomainTransferability

Create a value of CheckDomainTransferability with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authCode:CheckDomainTransferability', checkDomainTransferability_authCode - If the registrar for the top-level domain (TLD) requires an authorization code to transfer the domain, the code that you got from the current registrar for the domain.

$sel:domainName:CheckDomainTransferability', checkDomainTransferability_domainName - The name of the domain that you want to transfer to Route 53. The top-level domain (TLD), such as .com, must be a TLD that Route 53 supports. For a list of supported TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

The domain name can contain only the following characters:

Letters a through z. Domain names are not case sensitive.

Numbers 0 through 9.

Hyphen (-). You can't specify a hyphen at the beginning or end of a label.

Period (.) to separate the labels in the name, such as the . in example.com.

Request Lenses

checkDomainTransferability_authCode :: Lens' CheckDomainTransferability (Maybe Text)

If the registrar for the top-level domain (TLD) requires an authorization code to transfer the domain, the code that you got from the current registrar for the domain.

checkDomainTransferability_domainName :: Lens' CheckDomainTransferability Text

The name of the domain that you want to transfer to Route 53. The top-level domain (TLD), such as .com, must be a TLD that Route 53 supports. For a list of supported TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

The domain name can contain only the following characters:

Letters a through z. Domain names are not case sensitive.

Numbers 0 through 9.

Hyphen (-). You can't specify a hyphen at the beginning or end of a label.

Period (.) to separate the labels in the name, such as the . in example.com.

Destructuring the Response

data CheckDomainTransferabilityResponse

The CheckDomainTransferability response includes the following elements.

See: newCheckDomainTransferabilityResponse smart constructor.

Constructors

CheckDomainTransferabilityResponse'

Fields

httpStatus :: Int

The response's http status code.

transferability :: DomainTransferability

A complex type that contains information about whether the specified domain can be transferred to Route 53.

newCheckDomainTransferabilityResponse

Arguments

:: Int
$sel:httpStatus:CheckDomainTransferabilityResponse'

-> DomainTransferability
$sel:transferability:CheckDomainTransferabilityResponse'

-> CheckDomainTransferabilityResponse

Create a value of CheckDomainTransferabilityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CheckDomainTransferabilityResponse', checkDomainTransferabilityResponse_httpStatus - The response's http status code.

$sel:transferability:CheckDomainTransferabilityResponse', checkDomainTransferabilityResponse_transferability - A complex type that contains information about whether the specified domain can be transferred to Route 53.

Response Lenses

checkDomainTransferabilityResponse_httpStatus :: Lens' CheckDomainTransferabilityResponse Int

The response's http status code.

checkDomainTransferabilityResponse_transferability :: Lens' CheckDomainTransferabilityResponse DomainTransferability

A complex type that contains information about whether the specified domain can be transferred to Route 53.