Network.AWS.Route53Domains.DisableDomainTransferLock

Description

This operation removes the transfer lock on the domain (specifically the clientTransferProhibited status) to allow domain transfers. We recommend you refrain from performing this action unless you intend to transfer the domain to a different registrar. Successful submission returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

Creating a Request

data DisableDomainTransferLock

The DisableDomainTransferLock request includes the following element.

See: newDisableDomainTransferLock smart constructor.

Constructors

DisableDomainTransferLock'

Fields

domainName :: Text

The name of the domain that you want to remove the transfer lock for.

newDisableDomainTransferLock

Arguments

:: Text
$sel:domainName:DisableDomainTransferLock'

-> DisableDomainTransferLock

Create a value of DisableDomainTransferLock with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:DisableDomainTransferLock', disableDomainTransferLock_domainName - The name of the domain that you want to remove the transfer lock for.

Request Lenses

disableDomainTransferLock_domainName :: Lens' DisableDomainTransferLock Text

The name of the domain that you want to remove the transfer lock for.

Destructuring the Response

data DisableDomainTransferLockResponse

The DisableDomainTransferLock response includes the following element.

See: newDisableDomainTransferLockResponse smart constructor.

Constructors

DisableDomainTransferLockResponse'

Fields

httpStatus :: Int

The response's http status code.

operationId :: Text

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

newDisableDomainTransferLockResponse

Arguments

:: Int
$sel:httpStatus:DisableDomainTransferLockResponse'

-> Text
$sel:operationId:DisableDomainTransferLockResponse'

-> DisableDomainTransferLockResponse

Create a value of DisableDomainTransferLockResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisableDomainTransferLockResponse', disableDomainTransferLockResponse_httpStatus - The response's http status code.

$sel:operationId:DisableDomainTransferLockResponse', disableDomainTransferLockResponse_operationId - Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

Response Lenses

disableDomainTransferLockResponse_httpStatus :: Lens' DisableDomainTransferLockResponse Int

The response's http status code.

disableDomainTransferLockResponse_operationId :: Lens' DisableDomainTransferLockResponse Text

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.