Network.AWS.Route53Domains.EnableDomainTransferLock

Description

This operation sets the transfer lock on the domain (specifically the clientTransferProhibited status) to prevent domain transfers. Successful submission returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

Creating a Request

data EnableDomainTransferLock

A request to set the transfer lock for the specified domain.

See: newEnableDomainTransferLock smart constructor.

Constructors

EnableDomainTransferLock'

Fields

domainName :: Text

The name of the domain that you want to set the transfer lock for.

newEnableDomainTransferLock

Arguments

:: Text
$sel:domainName:EnableDomainTransferLock'

-> EnableDomainTransferLock

Create a value of EnableDomainTransferLock with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:EnableDomainTransferLock', enableDomainTransferLock_domainName - The name of the domain that you want to set the transfer lock for.

Request Lenses

enableDomainTransferLock_domainName :: Lens' EnableDomainTransferLock Text

The name of the domain that you want to set the transfer lock for.

Destructuring the Response

data EnableDomainTransferLockResponse

The EnableDomainTransferLock response includes the following elements.

See: newEnableDomainTransferLockResponse smart constructor.

Constructors

EnableDomainTransferLockResponse'

Fields

httpStatus :: Int

The response's http status code.

operationId :: Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail.

newEnableDomainTransferLockResponse

Arguments

:: Int
$sel:httpStatus:EnableDomainTransferLockResponse'

-> Text
$sel:operationId:EnableDomainTransferLockResponse'

-> EnableDomainTransferLockResponse

Create a value of EnableDomainTransferLockResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:EnableDomainTransferLockResponse', enableDomainTransferLockResponse_httpStatus - The response's http status code.

$sel:operationId:EnableDomainTransferLockResponse', enableDomainTransferLockResponse_operationId - Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail.

Response Lenses

enableDomainTransferLockResponse_httpStatus :: Lens' EnableDomainTransferLockResponse Int

The response's http status code.

enableDomainTransferLockResponse_operationId :: Lens' EnableDomainTransferLockResponse Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail.