Network.AWS.Route53Domains.GetDomainDetail

Description

This operation returns detailed information about a specified domain that is associated with the current AWS account. Contact information for the domain is also returned as part of the output.

Creating a Request

data GetDomainDetail

The GetDomainDetail request includes the following element.

See: newGetDomainDetail smart constructor.

Constructors

GetDomainDetail'

Fields

domainName :: Text

The name of the domain that you want to get detailed information about.

newGetDomainDetail

Arguments

:: Text
$sel:domainName:GetDomainDetail'

-> GetDomainDetail

Create a value of GetDomainDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:GetDomainDetail', getDomainDetail_domainName - The name of the domain that you want to get detailed information about.

Request Lenses

getDomainDetail_domainName :: Lens' GetDomainDetail Text

The name of the domain that you want to get detailed information about.

Destructuring the Response

data GetDomainDetailResponse

The GetDomainDetail response includes the following elements.

See: newGetDomainDetailResponse smart constructor.

Constructors

GetDomainDetailResponse'

Fields

techPrivacy :: Maybe Bool

Specifies whether contact information is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If the value is false, WHOIS queries return the information that you entered for the technical contact.

dnsSec :: Maybe Text

Reserved for future use.

whoIsServer :: Maybe Text

The fully qualified name of the WHOIS server that can answer the WHOIS query for the domain.

registryDomainId :: Maybe Text

Reserved for future use.

registrantPrivacy :: Maybe Bool

Specifies whether contact information is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If the value is false, WHOIS queries return the information that you entered for the registrant contact (domain owner).

updatedDate :: Maybe POSIX

The last updated date of the domain as found in the response to a WHOIS query. The date and time is in Unix time format and Coordinated Universal time (UTC).

adminPrivacy :: Maybe Bool

Specifies whether contact information is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If the value is false, WHOIS queries return the information that you entered for the admin contact.

autoRenew :: Maybe Bool

Specifies whether the domain registration is set to renew automatically.

abuseContactPhone :: Maybe Text

Phone number for reporting abuse.

registrarUrl :: Maybe Text

Web address of the registrar.

abuseContactEmail :: Maybe Text

Email address to contact to report incorrect contact information for a domain, to report that the domain is being used to send spam, to report that someone is cybersquatting on a domain name, or report some other type of abuse.

expirationDate :: Maybe POSIX

The date when the registration for the domain is set to expire. The date and time is in Unix time format and Coordinated Universal time (UTC).

creationDate :: Maybe POSIX

The date when the domain was created as found in the response to a WHOIS query. The date and time is in Unix time format and Coordinated Universal time (UTC).

registrarName :: Maybe Text

Name of the registrar of the domain as identified in the registry. Domains with a .com, .net, or .org TLD are registered by Amazon Registrar. All other domains are registered by our registrar associate, Gandi. The value for domains that are registered by Gandi is "GANDI SAS".

reseller :: Maybe Text

Reseller of the domain. Domains registered or transferred using Route 53 domains will have "Amazon" as the reseller.

statusList :: Maybe [Text]

An array of domain name status codes, also known as Extensible Provisioning Protocol (EPP) status codes.

ICANN, the organization that maintains a central database of domain names, has developed a set of domain name status codes that tell you the status of a variety of operations on a domain name, for example, registering a domain name, transferring a domain name to another registrar, renewing the registration for a domain name, and so on. All registrars use this same set of status codes.

For a current list of domain name status codes and an explanation of what each code means, go to the ICANN website and search for epp status codes. (Search on the ICANN website; web searches sometimes return an old version of the document.)

httpStatus :: Int

The response's http status code.

domainName :: Text

The name of a domain.

nameservers :: [Nameserver]

The name of the domain.

adminContact :: Sensitive ContactDetail

Provides details about the domain administrative contact.

registrantContact :: Sensitive ContactDetail

Provides details about the domain registrant.

techContact :: Sensitive ContactDetail

Provides details about the domain technical contact.

newGetDomainDetailResponse

Arguments

:: Int
$sel:httpStatus:GetDomainDetailResponse'

-> Text
$sel:domainName:GetDomainDetail'

-> ContactDetail
$sel:adminContact:GetDomainDetailResponse'

-> ContactDetail
$sel:registrantContact:GetDomainDetailResponse'

-> ContactDetail
$sel:techContact:GetDomainDetailResponse'

-> GetDomainDetailResponse

Create a value of GetDomainDetailResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:techPrivacy:GetDomainDetailResponse', getDomainDetailResponse_techPrivacy - Specifies whether contact information is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If the value is false, WHOIS queries return the information that you entered for the technical contact.

$sel:dnsSec:GetDomainDetailResponse', getDomainDetailResponse_dnsSec - Reserved for future use.

$sel:whoIsServer:GetDomainDetailResponse', getDomainDetailResponse_whoIsServer - The fully qualified name of the WHOIS server that can answer the WHOIS query for the domain.

$sel:registryDomainId:GetDomainDetailResponse', getDomainDetailResponse_registryDomainId - Reserved for future use.

$sel:registrantPrivacy:GetDomainDetailResponse', getDomainDetailResponse_registrantPrivacy - Specifies whether contact information is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If the value is false, WHOIS queries return the information that you entered for the registrant contact (domain owner).

$sel:updatedDate:GetDomainDetailResponse', getDomainDetailResponse_updatedDate - The last updated date of the domain as found in the response to a WHOIS query. The date and time is in Unix time format and Coordinated Universal time (UTC).

$sel:adminPrivacy:GetDomainDetailResponse', getDomainDetailResponse_adminPrivacy - Specifies whether contact information is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If the value is false, WHOIS queries return the information that you entered for the admin contact.

$sel:autoRenew:GetDomainDetailResponse', getDomainDetailResponse_autoRenew - Specifies whether the domain registration is set to renew automatically.

$sel:abuseContactPhone:GetDomainDetailResponse', getDomainDetailResponse_abuseContactPhone - Phone number for reporting abuse.

$sel:registrarUrl:GetDomainDetailResponse', getDomainDetailResponse_registrarUrl - Web address of the registrar.

$sel:abuseContactEmail:GetDomainDetailResponse', getDomainDetailResponse_abuseContactEmail - Email address to contact to report incorrect contact information for a domain, to report that the domain is being used to send spam, to report that someone is cybersquatting on a domain name, or report some other type of abuse.

$sel:expirationDate:GetDomainDetailResponse', getDomainDetailResponse_expirationDate - The date when the registration for the domain is set to expire. The date and time is in Unix time format and Coordinated Universal time (UTC).

$sel:creationDate:GetDomainDetailResponse', getDomainDetailResponse_creationDate - The date when the domain was created as found in the response to a WHOIS query. The date and time is in Unix time format and Coordinated Universal time (UTC).

$sel:registrarName:GetDomainDetailResponse', getDomainDetailResponse_registrarName - Name of the registrar of the domain as identified in the registry. Domains with a .com, .net, or .org TLD are registered by Amazon Registrar. All other domains are registered by our registrar associate, Gandi. The value for domains that are registered by Gandi is "GANDI SAS".

$sel:reseller:GetDomainDetailResponse', getDomainDetailResponse_reseller - Reseller of the domain. Domains registered or transferred using Route 53 domains will have "Amazon" as the reseller.

$sel:statusList:GetDomainDetailResponse', getDomainDetailResponse_statusList - An array of domain name status codes, also known as Extensible Provisioning Protocol (EPP) status codes.

ICANN, the organization that maintains a central database of domain names, has developed a set of domain name status codes that tell you the status of a variety of operations on a domain name, for example, registering a domain name, transferring a domain name to another registrar, renewing the registration for a domain name, and so on. All registrars use this same set of status codes.

For a current list of domain name status codes and an explanation of what each code means, go to the ICANN website and search for epp status codes. (Search on the ICANN website; web searches sometimes return an old version of the document.)

$sel:httpStatus:GetDomainDetailResponse', getDomainDetailResponse_httpStatus - The response's http status code.

$sel:domainName:GetDomainDetail', getDomainDetailResponse_domainName - The name of a domain.

$sel:nameservers:GetDomainDetailResponse', getDomainDetailResponse_nameservers - The name of the domain.

$sel:adminContact:GetDomainDetailResponse', getDomainDetailResponse_adminContact - Provides details about the domain administrative contact.

$sel:registrantContact:GetDomainDetailResponse', getDomainDetailResponse_registrantContact - Provides details about the domain registrant.

$sel:techContact:GetDomainDetailResponse', getDomainDetailResponse_techContact - Provides details about the domain technical contact.

Response Lenses

getDomainDetailResponse_techPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether contact information is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If the value is false, WHOIS queries return the information that you entered for the technical contact.

getDomainDetailResponse_dnsSec :: Lens' GetDomainDetailResponse (Maybe Text)

Reserved for future use.

getDomainDetailResponse_whoIsServer :: Lens' GetDomainDetailResponse (Maybe Text)

The fully qualified name of the WHOIS server that can answer the WHOIS query for the domain.

getDomainDetailResponse_registryDomainId :: Lens' GetDomainDetailResponse (Maybe Text)

Reserved for future use.

getDomainDetailResponse_registrantPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether contact information is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If the value is false, WHOIS queries return the information that you entered for the registrant contact (domain owner).

getDomainDetailResponse_updatedDate :: Lens' GetDomainDetailResponse (Maybe UTCTime)

The last updated date of the domain as found in the response to a WHOIS query. The date and time is in Unix time format and Coordinated Universal time (UTC).

getDomainDetailResponse_adminPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether contact information is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If the value is false, WHOIS queries return the information that you entered for the admin contact.

getDomainDetailResponse_autoRenew :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether the domain registration is set to renew automatically.

getDomainDetailResponse_abuseContactPhone :: Lens' GetDomainDetailResponse (Maybe Text)

Phone number for reporting abuse.

getDomainDetailResponse_registrarUrl :: Lens' GetDomainDetailResponse (Maybe Text)

Web address of the registrar.

getDomainDetailResponse_abuseContactEmail :: Lens' GetDomainDetailResponse (Maybe Text)

Email address to contact to report incorrect contact information for a domain, to report that the domain is being used to send spam, to report that someone is cybersquatting on a domain name, or report some other type of abuse.

getDomainDetailResponse_expirationDate :: Lens' GetDomainDetailResponse (Maybe UTCTime)

The date when the registration for the domain is set to expire. The date and time is in Unix time format and Coordinated Universal time (UTC).

getDomainDetailResponse_creationDate :: Lens' GetDomainDetailResponse (Maybe UTCTime)

The date when the domain was created as found in the response to a WHOIS query. The date and time is in Unix time format and Coordinated Universal time (UTC).

getDomainDetailResponse_registrarName :: Lens' GetDomainDetailResponse (Maybe Text)

Name of the registrar of the domain as identified in the registry. Domains with a .com, .net, or .org TLD are registered by Amazon Registrar. All other domains are registered by our registrar associate, Gandi. The value for domains that are registered by Gandi is "GANDI SAS".

getDomainDetailResponse_reseller :: Lens' GetDomainDetailResponse (Maybe Text)

Reseller of the domain. Domains registered or transferred using Route 53 domains will have "Amazon" as the reseller.

getDomainDetailResponse_statusList :: Lens' GetDomainDetailResponse (Maybe [Text])

An array of domain name status codes, also known as Extensible Provisioning Protocol (EPP) status codes.

ICANN, the organization that maintains a central database of domain names, has developed a set of domain name status codes that tell you the status of a variety of operations on a domain name, for example, registering a domain name, transferring a domain name to another registrar, renewing the registration for a domain name, and so on. All registrars use this same set of status codes.

For a current list of domain name status codes and an explanation of what each code means, go to the ICANN website and search for epp status codes. (Search on the ICANN website; web searches sometimes return an old version of the document.)

getDomainDetailResponse_httpStatus :: Lens' GetDomainDetailResponse Int

The response's http status code.

getDomainDetailResponse_domainName :: Lens' GetDomainDetailResponse Text

The name of a domain.

getDomainDetailResponse_nameservers :: Lens' GetDomainDetailResponse [Nameserver]

The name of the domain.

getDomainDetailResponse_adminContact :: Lens' GetDomainDetailResponse ContactDetail

Provides details about the domain administrative contact.

getDomainDetailResponse_registrantContact :: Lens' GetDomainDetailResponse ContactDetail

Provides details about the domain registrant.

getDomainDetailResponse_techContact :: Lens' GetDomainDetailResponse ContactDetail

Provides details about the domain technical contact.