Network.AWS.Route53Domains.GetDomainSuggestions

Description

The GetDomainSuggestions operation returns a list of suggested domain names.

Creating a Request

data GetDomainSuggestions

See: newGetDomainSuggestions smart constructor.

Constructors

GetDomainSuggestions'

Fields

domainName :: Text

A domain name that you want to use as the basis for a list of possible domain names. The top-level domain (TLD), such as .com, must be a TLD that Route 53 supports. For a list of supported TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

The domain name can contain only the following characters:

Letters a through z. Domain names are not case sensitive.

Numbers 0 through 9.

Hyphen (-). You can't specify a hyphen at the beginning or end of a label.

Period (.) to separate the labels in the name, such as the . in example.com.

Internationalized domain names are not supported for some top-level domains. To determine whether the TLD that you want to use supports internationalized domain names, see Domains that You Can Register with Amazon Route 53.

suggestionCount :: Int

The number of suggested domain names that you want Route 53 to return. Specify a value between 1 and 50.

onlyAvailable :: Bool

If OnlyAvailable is true, Route 53 returns only domain names that are available. If OnlyAvailable is false, Route 53 returns domain names without checking whether they're available to be registered. To determine whether the domain is available, you can call checkDomainAvailability for each suggestion.

newGetDomainSuggestions

Arguments

:: Text
$sel:domainName:GetDomainSuggestions'

-> Int
$sel:suggestionCount:GetDomainSuggestions'

-> Bool
$sel:onlyAvailable:GetDomainSuggestions'

-> GetDomainSuggestions

Create a value of GetDomainSuggestions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:GetDomainSuggestions', getDomainSuggestions_domainName - A domain name that you want to use as the basis for a list of possible domain names. The top-level domain (TLD), such as .com, must be a TLD that Route 53 supports. For a list of supported TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

The domain name can contain only the following characters:

Letters a through z. Domain names are not case sensitive.

Numbers 0 through 9.

Hyphen (-). You can't specify a hyphen at the beginning or end of a label.

Period (.) to separate the labels in the name, such as the . in example.com.

Internationalized domain names are not supported for some top-level domains. To determine whether the TLD that you want to use supports internationalized domain names, see Domains that You Can Register with Amazon Route 53.

$sel:suggestionCount:GetDomainSuggestions', getDomainSuggestions_suggestionCount - The number of suggested domain names that you want Route 53 to return. Specify a value between 1 and 50.

$sel:onlyAvailable:GetDomainSuggestions', getDomainSuggestions_onlyAvailable - If OnlyAvailable is true, Route 53 returns only domain names that are available. If OnlyAvailable is false, Route 53 returns domain names without checking whether they're available to be registered. To determine whether the domain is available, you can call checkDomainAvailability for each suggestion.

Request Lenses

getDomainSuggestions_domainName :: Lens' GetDomainSuggestions Text

A domain name that you want to use as the basis for a list of possible domain names. The top-level domain (TLD), such as .com, must be a TLD that Route 53 supports. For a list of supported TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

The domain name can contain only the following characters:

Letters a through z. Domain names are not case sensitive.

Numbers 0 through 9.

Hyphen (-). You can't specify a hyphen at the beginning or end of a label.

Period (.) to separate the labels in the name, such as the . in example.com.

Internationalized domain names are not supported for some top-level domains. To determine whether the TLD that you want to use supports internationalized domain names, see Domains that You Can Register with Amazon Route 53.

getDomainSuggestions_suggestionCount :: Lens' GetDomainSuggestions Int

The number of suggested domain names that you want Route 53 to return. Specify a value between 1 and 50.

getDomainSuggestions_onlyAvailable :: Lens' GetDomainSuggestions Bool

If OnlyAvailable is true, Route 53 returns only domain names that are available. If OnlyAvailable is false, Route 53 returns domain names without checking whether they're available to be registered. To determine whether the domain is available, you can call checkDomainAvailability for each suggestion.

Destructuring the Response

data GetDomainSuggestionsResponse

See: newGetDomainSuggestionsResponse smart constructor.

Constructors

GetDomainSuggestionsResponse'

Fields

suggestionsList :: Maybe [DomainSuggestion]

A list of possible domain names. If you specified true for OnlyAvailable in the request, the list contains only domains that are available for registration.

httpStatus :: Int

The response's http status code.

newGetDomainSuggestionsResponse

Arguments

:: Int
$sel:httpStatus:GetDomainSuggestionsResponse'

-> GetDomainSuggestionsResponse

Create a value of GetDomainSuggestionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suggestionsList:GetDomainSuggestionsResponse', getDomainSuggestionsResponse_suggestionsList - A list of possible domain names. If you specified true for OnlyAvailable in the request, the list contains only domains that are available for registration.

$sel:httpStatus:GetDomainSuggestionsResponse', getDomainSuggestionsResponse_httpStatus - The response's http status code.

Response Lenses

getDomainSuggestionsResponse_suggestionsList :: Lens' GetDomainSuggestionsResponse (Maybe [DomainSuggestion])

A list of possible domain names. If you specified true for OnlyAvailable in the request, the list contains only domains that are available for registration.

getDomainSuggestionsResponse_httpStatus :: Lens' GetDomainSuggestionsResponse Int

The response's http status code.