Network.AWS.Route53Domains.GetOperationDetail

Description

This operation returns the current status of an operation that is not completed.

Creating a Request

data GetOperationDetail

The GetOperationDetail request includes the following element.

See: newGetOperationDetail smart constructor.

Constructors

GetOperationDetail'

Fields

operationId :: Text

The identifier for the operation for which you want to get the status. Route 53 returned the identifier in the response to the original request.

newGetOperationDetail

Arguments

:: Text
$sel:operationId:GetOperationDetail'

-> GetOperationDetail

Create a value of GetOperationDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationId:GetOperationDetail', getOperationDetail_operationId - The identifier for the operation for which you want to get the status. Route 53 returned the identifier in the response to the original request.

Request Lenses

getOperationDetail_operationId :: Lens' GetOperationDetail Text

The identifier for the operation for which you want to get the status. Route 53 returned the identifier in the response to the original request.

Destructuring the Response

data GetOperationDetailResponse

The GetOperationDetail response includes the following elements.

See: newGetOperationDetailResponse smart constructor.

Constructors

GetOperationDetailResponse'

Fields

status :: Maybe OperationStatus

The current status of the requested operation in the system.

message :: Maybe Text

Detailed information on the status including possible errors.

operationId :: Maybe Text

The identifier for the operation.

submittedDate :: Maybe POSIX

The date when the request was submitted.

domainName :: Maybe Text

The name of a domain.

type' :: Maybe OperationType

The type of operation that was requested.

httpStatus :: Int

The response's http status code.

newGetOperationDetailResponse

Arguments

:: Int
$sel:httpStatus:GetOperationDetailResponse'

-> GetOperationDetailResponse

Create a value of GetOperationDetailResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetOperationDetailResponse', getOperationDetailResponse_status - The current status of the requested operation in the system.

$sel:message:GetOperationDetailResponse', getOperationDetailResponse_message - Detailed information on the status including possible errors.

$sel:operationId:GetOperationDetail', getOperationDetailResponse_operationId - The identifier for the operation.

$sel:submittedDate:GetOperationDetailResponse', getOperationDetailResponse_submittedDate - The date when the request was submitted.

$sel:domainName:GetOperationDetailResponse', getOperationDetailResponse_domainName - The name of a domain.

$sel:type':GetOperationDetailResponse', getOperationDetailResponse_type - The type of operation that was requested.

$sel:httpStatus:GetOperationDetailResponse', getOperationDetailResponse_httpStatus - The response's http status code.

Response Lenses

getOperationDetailResponse_status :: Lens' GetOperationDetailResponse (Maybe OperationStatus)

The current status of the requested operation in the system.

getOperationDetailResponse_message :: Lens' GetOperationDetailResponse (Maybe Text)

Detailed information on the status including possible errors.

getOperationDetailResponse_operationId :: Lens' GetOperationDetailResponse (Maybe Text)

The identifier for the operation.

getOperationDetailResponse_submittedDate :: Lens' GetOperationDetailResponse (Maybe UTCTime)

The date when the request was submitted.

getOperationDetailResponse_domainName :: Lens' GetOperationDetailResponse (Maybe Text)

The name of a domain.

getOperationDetailResponse_type :: Lens' GetOperationDetailResponse (Maybe OperationType)

The type of operation that was requested.

getOperationDetailResponse_httpStatus :: Lens' GetOperationDetailResponse Int

The response's http status code.