Network.AWS.Route53Domains.Lens

Description

Operations

CheckDomainTransferability

checkDomainTransferability_authCode :: Lens' CheckDomainTransferability (Maybe Text)

If the registrar for the top-level domain (TLD) requires an authorization code to transfer the domain, the code that you got from the current registrar for the domain.

checkDomainTransferability_domainName :: Lens' CheckDomainTransferability Text

The name of the domain that you want to transfer to Route 53. The top-level domain (TLD), such as .com, must be a TLD that Route 53 supports. For a list of supported TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

The domain name can contain only the following characters:

Letters a through z. Domain names are not case sensitive.

Numbers 0 through 9.

Hyphen (-). You can't specify a hyphen at the beginning or end of a label.

Period (.) to separate the labels in the name, such as the . in example.com.

checkDomainTransferabilityResponse_httpStatus :: Lens' CheckDomainTransferabilityResponse Int

The response's http status code.

checkDomainTransferabilityResponse_transferability :: Lens' CheckDomainTransferabilityResponse DomainTransferability

A complex type that contains information about whether the specified domain can be transferred to Route 53.

CheckDomainAvailability

checkDomainAvailability_idnLangCode :: Lens' CheckDomainAvailability (Maybe Text)

Reserved for future use.

checkDomainAvailability_domainName :: Lens' CheckDomainAvailability Text

The name of the domain that you want to get availability for. The top-level domain (TLD), such as .com, must be a TLD that Route 53 supports. For a list of supported TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

The domain name can contain only the following characters:

Letters a through z. Domain names are not case sensitive.

Numbers 0 through 9.

Hyphen (-). You can't specify a hyphen at the beginning or end of a label.

Period (.) to separate the labels in the name, such as the . in example.com.

Internationalized domain names are not supported for some top-level domains. To determine whether the TLD that you want to use supports internationalized domain names, see Domains that You Can Register with Amazon Route 53. For more information, see Formatting Internationalized Domain Names.

checkDomainAvailabilityResponse_httpStatus :: Lens' CheckDomainAvailabilityResponse Int

The response's http status code.

checkDomainAvailabilityResponse_availability :: Lens' CheckDomainAvailabilityResponse DomainAvailability

Whether the domain name is available for registering.

You can register only domains designated as AVAILABLE.

Valid values:

AVAILABLE

The domain name is available.

AVAILABLE_RESERVED

The domain name is reserved under specific conditions.

AVAILABLE_PREORDER

The domain name is available and can be preordered.

DONT_KNOW

The TLD registry didn't reply with a definitive answer about whether the domain name is available. Route 53 can return this response for a variety of reasons, for example, the registry is performing maintenance. Try again later.

PENDING

The TLD registry didn't return a response in the expected amount of time. When the response is delayed, it usually takes just a few extra seconds. You can resubmit the request immediately.

RESERVED

The domain name has been reserved for another person or organization.

UNAVAILABLE

The domain name is not available.

UNAVAILABLE_PREMIUM

The domain name is not available.

UNAVAILABLE_RESTRICTED

The domain name is forbidden.

TransferDomainToAnotherAwsAccount

transferDomainToAnotherAwsAccount_domainName :: Lens' TransferDomainToAnotherAwsAccount Text

The name of the domain that you want to transfer from the current AWS account to another account.

transferDomainToAnotherAwsAccount_accountId :: Lens' TransferDomainToAnotherAwsAccount Text

The account ID of the AWS account that you want to transfer the domain to, for example, 111122223333.

transferDomainToAnotherAwsAccountResponse_operationId :: Lens' TransferDomainToAnotherAwsAccountResponse (Maybe Text)

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

transferDomainToAnotherAwsAccountResponse_password :: Lens' TransferDomainToAnotherAwsAccountResponse (Maybe Text)

To finish transferring a domain to another AWS account, the account that the domain is being transferred to must submit an AcceptDomainTransferFromAnotherAwsAccount request. The request must include the value of the Password element that was returned in the TransferDomainToAnotherAwsAccount response.

transferDomainToAnotherAwsAccountResponse_httpStatus :: Lens' TransferDomainToAnotherAwsAccountResponse Int

The response's http status code.

UpdateDomainContactPrivacy

updateDomainContactPrivacy_adminPrivacy :: Lens' UpdateDomainContactPrivacy (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the admin contact.

updateDomainContactPrivacy_techPrivacy :: Lens' UpdateDomainContactPrivacy (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the technical contact.

updateDomainContactPrivacy_registrantPrivacy :: Lens' UpdateDomainContactPrivacy (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the registrant contact (domain owner).

updateDomainContactPrivacy_domainName :: Lens' UpdateDomainContactPrivacy Text

The name of the domain that you want to update the privacy setting for.

updateDomainContactPrivacyResponse_httpStatus :: Lens' UpdateDomainContactPrivacyResponse Int

The response's http status code.

updateDomainContactPrivacyResponse_operationId :: Lens' UpdateDomainContactPrivacyResponse Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail.

ListOperations

listOperations_submittedSince :: Lens' ListOperations (Maybe UTCTime)

An optional parameter that lets you get information about all the operations that you submitted after a specified date and time. Specify the date and time in Unix time format and Coordinated Universal time (UTC).

listOperations_maxItems :: Lens' ListOperations (Maybe Int)

Number of domains to be returned.

Default: 20

listOperations_marker :: Lens' ListOperations (Maybe Text)

For an initial request for a list of operations, omit this element. If the number of operations that are not yet complete is greater than the value that you specified for MaxItems, you can use Marker to return additional operations. Get the value of NextPageMarker from the previous response, and submit another request that includes the value of NextPageMarker in the Marker element.

listOperationsResponse_nextPageMarker :: Lens' ListOperationsResponse (Maybe Text)

If there are more operations than you specified for MaxItems in the request, submit another request and include the value of NextPageMarker in the value of Marker.

listOperationsResponse_httpStatus :: Lens' ListOperationsResponse Int

The response's http status code.

listOperationsResponse_operations :: Lens' ListOperationsResponse [OperationSummary]

Lists summaries of the operations.

DisableDomainTransferLock

disableDomainTransferLock_domainName :: Lens' DisableDomainTransferLock Text

The name of the domain that you want to remove the transfer lock for.

disableDomainTransferLockResponse_httpStatus :: Lens' DisableDomainTransferLockResponse Int

The response's http status code.

disableDomainTransferLockResponse_operationId :: Lens' DisableDomainTransferLockResponse Text

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

EnableDomainTransferLock

enableDomainTransferLock_domainName :: Lens' EnableDomainTransferLock Text

The name of the domain that you want to set the transfer lock for.

enableDomainTransferLockResponse_httpStatus :: Lens' EnableDomainTransferLockResponse Int

The response's http status code.

enableDomainTransferLockResponse_operationId :: Lens' EnableDomainTransferLockResponse Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail.

RegisterDomain

registerDomain_autoRenew :: Lens' RegisterDomain (Maybe Bool)

Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged.

Default: true

registerDomain_idnLangCode :: Lens' RegisterDomain (Maybe Text)

Reserved for future use.

registerDomain_privacyProtectTechContact :: Lens' RegisterDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the technical contact.

Default: true

registerDomain_privacyProtectAdminContact :: Lens' RegisterDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the admin contact.

Default: true

registerDomain_privacyProtectRegistrantContact :: Lens' RegisterDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the registrant contact (the domain owner).

Default: true

registerDomain_domainName :: Lens' RegisterDomain Text

The domain name that you want to register. The top-level domain (TLD), such as .com, must be a TLD that Route 53 supports. For a list of supported TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

The domain name can contain only the following characters:

Letters a through z. Domain names are not case sensitive.

Numbers 0 through 9.

Hyphen (-). You can't specify a hyphen at the beginning or end of a label.

Period (.) to separate the labels in the name, such as the . in example.com.

Internationalized domain names are not supported for some top-level domains. To determine whether the TLD that you want to use supports internationalized domain names, see Domains that You Can Register with Amazon Route 53. For more information, see Formatting Internationalized Domain Names.

registerDomain_durationInYears :: Lens' RegisterDomain Natural

The number of years that you want to register the domain for. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain. For the range of valid values for your domain, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

Default: 1

registerDomain_adminContact :: Lens' RegisterDomain ContactDetail

Provides detailed contact information. For information about the values that you specify for each element, see ContactDetail.

registerDomain_registrantContact :: Lens' RegisterDomain ContactDetail

Provides detailed contact information. For information about the values that you specify for each element, see ContactDetail.

registerDomain_techContact :: Lens' RegisterDomain ContactDetail

Provides detailed contact information. For information about the values that you specify for each element, see ContactDetail.

registerDomainResponse_httpStatus :: Lens' RegisterDomainResponse Int

The response's http status code.

registerDomainResponse_operationId :: Lens' RegisterDomainResponse Text

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

GetDomainSuggestions

getDomainSuggestions_domainName :: Lens' GetDomainSuggestions Text

A domain name that you want to use as the basis for a list of possible domain names. The top-level domain (TLD), such as .com, must be a TLD that Route 53 supports. For a list of supported TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

The domain name can contain only the following characters:

Letters a through z. Domain names are not case sensitive.

Numbers 0 through 9.

Hyphen (-). You can't specify a hyphen at the beginning or end of a label.

Period (.) to separate the labels in the name, such as the . in example.com.

Internationalized domain names are not supported for some top-level domains. To determine whether the TLD that you want to use supports internationalized domain names, see Domains that You Can Register with Amazon Route 53.

getDomainSuggestions_suggestionCount :: Lens' GetDomainSuggestions Int

The number of suggested domain names that you want Route 53 to return. Specify a value between 1 and 50.

getDomainSuggestions_onlyAvailable :: Lens' GetDomainSuggestions Bool

If OnlyAvailable is true, Route 53 returns only domain names that are available. If OnlyAvailable is false, Route 53 returns domain names without checking whether they're available to be registered. To determine whether the domain is available, you can call checkDomainAvailability for each suggestion.

getDomainSuggestionsResponse_suggestionsList :: Lens' GetDomainSuggestionsResponse (Maybe [DomainSuggestion])

A list of possible domain names. If you specified true for OnlyAvailable in the request, the list contains only domains that are available for registration.

getDomainSuggestionsResponse_httpStatus :: Lens' GetDomainSuggestionsResponse Int

The response's http status code.

CancelDomainTransferToAnotherAwsAccount

cancelDomainTransferToAnotherAwsAccount_domainName :: Lens' CancelDomainTransferToAnotherAwsAccount Text

The name of the domain for which you want to cancel the transfer to another AWS account.

cancelDomainTransferToAnotherAwsAccountResponse_operationId :: Lens' CancelDomainTransferToAnotherAwsAccountResponse (Maybe Text)

The identifier that TransferDomainToAnotherAwsAccount returned to track the progress of the request. Because the transfer request was canceled, the value is no longer valid, and you can't use GetOperationDetail to query the operation status.

cancelDomainTransferToAnotherAwsAccountResponse_httpStatus :: Lens' CancelDomainTransferToAnotherAwsAccountResponse Int

The response's http status code.

ListDomains

listDomains_maxItems :: Lens' ListDomains (Maybe Int)

Number of domains to be returned.

Default: 20

listDomains_marker :: Lens' ListDomains (Maybe Text)

For an initial request for a list of domains, omit this element. If the number of domains that are associated with the current AWS account is greater than the value that you specified for MaxItems, you can use Marker to return additional domains. Get the value of NextPageMarker from the previous response, and submit another request that includes the value of NextPageMarker in the Marker element.

Constraints: The marker must match the value specified in the previous request.

listDomainsResponse_nextPageMarker :: Lens' ListDomainsResponse (Maybe Text)

If there are more domains than you specified for MaxItems in the request, submit another request and include the value of NextPageMarker in the value of Marker.

listDomainsResponse_httpStatus :: Lens' ListDomainsResponse Int

The response's http status code.

listDomainsResponse_domains :: Lens' ListDomainsResponse [DomainSummary]

A summary of domains.

ViewBilling

viewBilling_end :: Lens' ViewBilling (Maybe UTCTime)

The end date and time for the time period for which you want a list of billing records. Specify the date and time in Unix time format and Coordinated Universal time (UTC).

viewBilling_start :: Lens' ViewBilling (Maybe UTCTime)

The beginning date and time for the time period for which you want a list of billing records. Specify the date and time in Unix time format and Coordinated Universal time (UTC).

viewBilling_maxItems :: Lens' ViewBilling (Maybe Int)

The number of billing records to be returned.

Default: 20

viewBilling_marker :: Lens' ViewBilling (Maybe Text)

For an initial request for a list of billing records, omit this element. If the number of billing records that are associated with the current AWS account during the specified period is greater than the value that you specified for MaxItems, you can use Marker to return additional billing records. Get the value of NextPageMarker from the previous response, and submit another request that includes the value of NextPageMarker in the Marker element.

Constraints: The marker must match the value of NextPageMarker that was returned in the previous response.

viewBillingResponse_billingRecords :: Lens' ViewBillingResponse (Maybe [BillingRecord])

A summary of billing records.

viewBillingResponse_nextPageMarker :: Lens' ViewBillingResponse (Maybe Text)

If there are more billing records than you specified for MaxItems in the request, submit another request and include the value of NextPageMarker in the value of Marker.

viewBillingResponse_httpStatus :: Lens' ViewBillingResponse Int

The response's http status code.

UpdateTagsForDomain

updateTagsForDomain_tagsToUpdate :: Lens' UpdateTagsForDomain (Maybe [Tag])

A list of the tag keys and values that you want to add or update. If you specify a key that already exists, the corresponding value will be replaced.

updateTagsForDomain_domainName :: Lens' UpdateTagsForDomain Text

The domain for which you want to add or update tags.

updateTagsForDomainResponse_httpStatus :: Lens' UpdateTagsForDomainResponse Int

The response's http status code.

DeleteTagsForDomain

deleteTagsForDomain_domainName :: Lens' DeleteTagsForDomain Text

The domain for which you want to delete one or more tags.

deleteTagsForDomain_tagsToDelete :: Lens' DeleteTagsForDomain [Text]

A list of tag keys to delete.

deleteTagsForDomainResponse_httpStatus :: Lens' DeleteTagsForDomainResponse Int

The response's http status code.

ListTagsForDomain

listTagsForDomain_domainName :: Lens' ListTagsForDomain Text

The domain for which you want to get a list of tags.

listTagsForDomainResponse_httpStatus :: Lens' ListTagsForDomainResponse Int

The response's http status code.

listTagsForDomainResponse_tagList :: Lens' ListTagsForDomainResponse [Tag]

A list of the tags that are associated with the specified domain.

ResendContactReachabilityEmail

resendContactReachabilityEmail_domainName :: Lens' ResendContactReachabilityEmail (Maybe Text)

The name of the domain for which you want Route 53 to resend a confirmation email to the registrant contact.

resendContactReachabilityEmailResponse_isAlreadyVerified :: Lens' ResendContactReachabilityEmailResponse (Maybe Bool)

True if the email address for the registrant contact has already been verified, and false otherwise. If the email address has already been verified, we don't send another confirmation email.

resendContactReachabilityEmailResponse_domainName :: Lens' ResendContactReachabilityEmailResponse (Maybe Text)

The domain name for which you requested a confirmation email.

resendContactReachabilityEmailResponse_emailAddress :: Lens' ResendContactReachabilityEmailResponse (Maybe Text)

The email address for the registrant contact at the time that we sent the verification email.

resendContactReachabilityEmailResponse_httpStatus :: Lens' ResendContactReachabilityEmailResponse Int

The response's http status code.

DisableDomainAutoRenew

disableDomainAutoRenew_domainName :: Lens' DisableDomainAutoRenew Text

The name of the domain that you want to disable automatic renewal for.

disableDomainAutoRenewResponse_httpStatus :: Lens' DisableDomainAutoRenewResponse Int

The response's http status code.

UpdateDomainNameservers

updateDomainNameservers_fIAuthKey :: Lens' UpdateDomainNameservers (Maybe Text)

The authorization key for .fi domains

updateDomainNameservers_domainName :: Lens' UpdateDomainNameservers Text

The name of the domain that you want to change name servers for.

updateDomainNameservers_nameservers :: Lens' UpdateDomainNameservers [Nameserver]

A list of new name servers for the domain.

updateDomainNameserversResponse_httpStatus :: Lens' UpdateDomainNameserversResponse Int

The response's http status code.

updateDomainNameserversResponse_operationId :: Lens' UpdateDomainNameserversResponse Text

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

EnableDomainAutoRenew

enableDomainAutoRenew_domainName :: Lens' EnableDomainAutoRenew Text

The name of the domain that you want to enable automatic renewal for.

enableDomainAutoRenewResponse_httpStatus :: Lens' EnableDomainAutoRenewResponse Int

The response's http status code.

GetContactReachabilityStatus

getContactReachabilityStatus_domainName :: Lens' GetContactReachabilityStatus (Maybe Text)

The name of the domain for which you want to know whether the registrant contact has confirmed that the email address is valid.

getContactReachabilityStatusResponse_status :: Lens' GetContactReachabilityStatusResponse (Maybe ReachabilityStatus)

Whether the registrant contact has responded. Values include the following:

PENDING

We sent the confirmation email and haven't received a response yet.

DONE

We sent the email and got confirmation from the registrant contact.

EXPIRED

The time limit expired before the registrant contact responded.

getContactReachabilityStatusResponse_domainName :: Lens' GetContactReachabilityStatusResponse (Maybe Text)

The domain name for which you requested the reachability status.

getContactReachabilityStatusResponse_httpStatus :: Lens' GetContactReachabilityStatusResponse Int

The response's http status code.

RejectDomainTransferFromAnotherAwsAccount

rejectDomainTransferFromAnotherAwsAccount_domainName :: Lens' RejectDomainTransferFromAnotherAwsAccount Text

The name of the domain that was specified when another AWS account submitted a TransferDomainToAnotherAwsAccount request.

rejectDomainTransferFromAnotherAwsAccountResponse_operationId :: Lens' RejectDomainTransferFromAnotherAwsAccountResponse (Maybe Text)

The identifier that TransferDomainToAnotherAwsAccount returned to track the progress of the request. Because the transfer request was rejected, the value is no longer valid, and you can't use GetOperationDetail to query the operation status.

rejectDomainTransferFromAnotherAwsAccountResponse_httpStatus :: Lens' RejectDomainTransferFromAnotherAwsAccountResponse Int

The response's http status code.

GetOperationDetail

getOperationDetail_operationId :: Lens' GetOperationDetail Text

The identifier for the operation for which you want to get the status. Route 53 returned the identifier in the response to the original request.

getOperationDetailResponse_status :: Lens' GetOperationDetailResponse (Maybe OperationStatus)

The current status of the requested operation in the system.

getOperationDetailResponse_message :: Lens' GetOperationDetailResponse (Maybe Text)

Detailed information on the status including possible errors.

getOperationDetailResponse_operationId :: Lens' GetOperationDetailResponse (Maybe Text)

The identifier for the operation.

getOperationDetailResponse_submittedDate :: Lens' GetOperationDetailResponse (Maybe UTCTime)

The date when the request was submitted.

getOperationDetailResponse_domainName :: Lens' GetOperationDetailResponse (Maybe Text)

The name of a domain.

getOperationDetailResponse_type :: Lens' GetOperationDetailResponse (Maybe OperationType)

The type of operation that was requested.

getOperationDetailResponse_httpStatus :: Lens' GetOperationDetailResponse Int

The response's http status code.

AcceptDomainTransferFromAnotherAwsAccount

acceptDomainTransferFromAnotherAwsAccount_domainName :: Lens' AcceptDomainTransferFromAnotherAwsAccount Text

The name of the domain that was specified when another AWS account submitted a TransferDomainToAnotherAwsAccount request.

acceptDomainTransferFromAnotherAwsAccount_password :: Lens' AcceptDomainTransferFromAnotherAwsAccount Text

The password that was returned by the TransferDomainToAnotherAwsAccount request.

acceptDomainTransferFromAnotherAwsAccountResponse_operationId :: Lens' AcceptDomainTransferFromAnotherAwsAccountResponse (Maybe Text)

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

acceptDomainTransferFromAnotherAwsAccountResponse_httpStatus :: Lens' AcceptDomainTransferFromAnotherAwsAccountResponse Int

The response's http status code.

GetDomainDetail

getDomainDetail_domainName :: Lens' GetDomainDetail Text

The name of the domain that you want to get detailed information about.

getDomainDetailResponse_dnsSec :: Lens' GetDomainDetailResponse (Maybe Text)

Reserved for future use.

getDomainDetailResponse_abuseContactEmail :: Lens' GetDomainDetailResponse (Maybe Text)

Email address to contact to report incorrect contact information for a domain, to report that the domain is being used to send spam, to report that someone is cybersquatting on a domain name, or report some other type of abuse.

getDomainDetailResponse_autoRenew :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether the domain registration is set to renew automatically.

getDomainDetailResponse_abuseContactPhone :: Lens' GetDomainDetailResponse (Maybe Text)

Phone number for reporting abuse.

getDomainDetailResponse_adminPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether contact information is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If the value is false, WHOIS queries return the information that you entered for the admin contact.

getDomainDetailResponse_statusList :: Lens' GetDomainDetailResponse (Maybe [Text])

An array of domain name status codes, also known as Extensible Provisioning Protocol (EPP) status codes.

ICANN, the organization that maintains a central database of domain names, has developed a set of domain name status codes that tell you the status of a variety of operations on a domain name, for example, registering a domain name, transferring a domain name to another registrar, renewing the registration for a domain name, and so on. All registrars use this same set of status codes.

For a current list of domain name status codes and an explanation of what each code means, go to the ICANN website and search for epp status codes. (Search on the ICANN website; web searches sometimes return an old version of the document.)

getDomainDetailResponse_reseller :: Lens' GetDomainDetailResponse (Maybe Text)

Reseller of the domain. Domains registered or transferred using Route 53 domains will have "Amazon" as the reseller.

getDomainDetailResponse_registryDomainId :: Lens' GetDomainDetailResponse (Maybe Text)

Reserved for future use.

getDomainDetailResponse_registrarName :: Lens' GetDomainDetailResponse (Maybe Text)

Name of the registrar of the domain as identified in the registry. Domains with a .com, .net, or .org TLD are registered by Amazon Registrar. All other domains are registered by our registrar associate, Gandi. The value for domains that are registered by Gandi is "GANDI SAS".

getDomainDetailResponse_creationDate :: Lens' GetDomainDetailResponse (Maybe UTCTime)

The date when the domain was created as found in the response to a WHOIS query. The date and time is in Unix time format and Coordinated Universal time (UTC).

getDomainDetailResponse_expirationDate :: Lens' GetDomainDetailResponse (Maybe UTCTime)

The date when the registration for the domain is set to expire. The date and time is in Unix time format and Coordinated Universal time (UTC).

getDomainDetailResponse_whoIsServer :: Lens' GetDomainDetailResponse (Maybe Text)

The fully qualified name of the WHOIS server that can answer the WHOIS query for the domain.

getDomainDetailResponse_registrarUrl :: Lens' GetDomainDetailResponse (Maybe Text)

Web address of the registrar.

getDomainDetailResponse_techPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether contact information is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If the value is false, WHOIS queries return the information that you entered for the technical contact.

getDomainDetailResponse_updatedDate :: Lens' GetDomainDetailResponse (Maybe UTCTime)

The last updated date of the domain as found in the response to a WHOIS query. The date and time is in Unix time format and Coordinated Universal time (UTC).

getDomainDetailResponse_registrantPrivacy :: Lens' GetDomainDetailResponse (Maybe Bool)

Specifies whether contact information is concealed from WHOIS queries. If the value is true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If the value is false, WHOIS queries return the information that you entered for the registrant contact (domain owner).

getDomainDetailResponse_httpStatus :: Lens' GetDomainDetailResponse Int

The response's http status code.

getDomainDetailResponse_domainName :: Lens' GetDomainDetailResponse Text

The name of a domain.

getDomainDetailResponse_nameservers :: Lens' GetDomainDetailResponse [Nameserver]

The name of the domain.

getDomainDetailResponse_adminContact :: Lens' GetDomainDetailResponse ContactDetail

Provides details about the domain administrative contact.

getDomainDetailResponse_registrantContact :: Lens' GetDomainDetailResponse ContactDetail

Provides details about the domain registrant.

getDomainDetailResponse_techContact :: Lens' GetDomainDetailResponse ContactDetail

Provides details about the domain technical contact.

UpdateDomainContact

updateDomainContact_registrantContact :: Lens' UpdateDomainContact (Maybe ContactDetail)

Provides detailed contact information.

updateDomainContact_techContact :: Lens' UpdateDomainContact (Maybe ContactDetail)

Provides detailed contact information.

updateDomainContact_adminContact :: Lens' UpdateDomainContact (Maybe ContactDetail)

Provides detailed contact information.

updateDomainContact_domainName :: Lens' UpdateDomainContact Text

The name of the domain that you want to update contact information for.

updateDomainContactResponse_httpStatus :: Lens' UpdateDomainContactResponse Int

The response's http status code.

updateDomainContactResponse_operationId :: Lens' UpdateDomainContactResponse Text

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

RenewDomain

renewDomain_durationInYears :: Lens' RenewDomain (Maybe Natural)

The number of years that you want to renew the domain for. The maximum number of years depends on the top-level domain. For the range of valid values for your domain, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

Default: 1

renewDomain_domainName :: Lens' RenewDomain Text

The name of the domain that you want to renew.

renewDomain_currentExpiryYear :: Lens' RenewDomain Int

The year when the registration for the domain is set to expire. This value must match the current expiration date for the domain.

renewDomainResponse_httpStatus :: Lens' RenewDomainResponse Int

The response's http status code.

renewDomainResponse_operationId :: Lens' RenewDomainResponse Text

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

TransferDomain

transferDomain_autoRenew :: Lens' TransferDomain (Maybe Bool)

Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged.

Default: true

transferDomain_nameservers :: Lens' TransferDomain (Maybe [Nameserver])

Contains details for the host and glue IP addresses.

transferDomain_authCode :: Lens' TransferDomain (Maybe Text)

The authorization code for the domain. You get this value from the current registrar.

transferDomain_idnLangCode :: Lens' TransferDomain (Maybe Text)

Reserved for future use.

transferDomain_privacyProtectTechContact :: Lens' TransferDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the technical contact.

Default: true

transferDomain_privacyProtectAdminContact :: Lens' TransferDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the admin contact.

Default: true

transferDomain_privacyProtectRegistrantContact :: Lens' TransferDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the registrant contact (domain owner).

Default: true

transferDomain_domainName :: Lens' TransferDomain Text

The name of the domain that you want to transfer to Route 53. The top-level domain (TLD), such as .com, must be a TLD that Route 53 supports. For a list of supported TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

The domain name can contain only the following characters:

Letters a through z. Domain names are not case sensitive.

Numbers 0 through 9.

Hyphen (-). You can't specify a hyphen at the beginning or end of a label.

Period (.) to separate the labels in the name, such as the . in example.com.

transferDomain_durationInYears :: Lens' TransferDomain Natural

The number of years that you want to register the domain for. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain.

Default: 1

transferDomain_adminContact :: Lens' TransferDomain ContactDetail

Provides detailed contact information.

transferDomain_registrantContact :: Lens' TransferDomain ContactDetail

Provides detailed contact information.

transferDomain_techContact :: Lens' TransferDomain ContactDetail

Provides detailed contact information.

transferDomainResponse_httpStatus :: Lens' TransferDomainResponse Int

The response's http status code.

transferDomainResponse_operationId :: Lens' TransferDomainResponse Text

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

RetrieveDomainAuthCode

retrieveDomainAuthCode_domainName :: Lens' RetrieveDomainAuthCode Text

The name of the domain that you want to get an authorization code for.

retrieveDomainAuthCodeResponse_httpStatus :: Lens' RetrieveDomainAuthCodeResponse Int

The response's http status code.

retrieveDomainAuthCodeResponse_authCode :: Lens' RetrieveDomainAuthCodeResponse Text

The authorization code for the domain.

Types

BillingRecord

billingRecord_invoiceId :: Lens' BillingRecord (Maybe Text)

The ID of the invoice that is associated with the billing record.

billingRecord_operation :: Lens' BillingRecord (Maybe OperationType)

The operation that you were charged for.

billingRecord_billDate :: Lens' BillingRecord (Maybe UTCTime)

The date that the operation was billed, in Unix format.

billingRecord_domainName :: Lens' BillingRecord (Maybe Text)

The name of the domain that the billing record applies to. If the domain name contains characters other than a-z, 0-9, and - (hyphen), such as an internationalized domain name, then this value is in Punycode. For more information, see DNS Domain Name Format in the Amazon Route 53 Developer Guide.

billingRecord_price :: Lens' BillingRecord (Maybe Double)

The price that you were charged for the operation, in US dollars.

Example value: 12.0

ContactDetail

contactDetail_phoneNumber :: Lens' ContactDetail (Maybe Text)

The phone number of the contact.

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code>]". For example, a US phone number might appear as "+1.1234567890".

contactDetail_organizationName :: Lens' ContactDetail (Maybe Text)

Name of the organization for contact types other than PERSON.

contactDetail_addressLine1 :: Lens' ContactDetail (Maybe Text)

First line of the contact's address.

contactDetail_extraParams :: Lens' ContactDetail (Maybe [ExtraParam])

A list of name-value pairs for parameters required by certain top-level domains.

contactDetail_zipCode :: Lens' ContactDetail (Maybe Text)

The zip or postal code of the contact's address.

contactDetail_contactType :: Lens' ContactDetail (Maybe ContactType)

Indicates whether the contact is a person, company, association, or public organization. Note the following:

If you specify a value other than PERSON, you must also specify a value for OrganizationName.

For some TLDs, the privacy protection available depends on the value that you specify for Contact Type. For the privacy protection settings for your TLD, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide

For .es domains, if you specify PERSON, you must specify INDIVIDUAL for the value of ES_LEGAL_FORM.

contactDetail_fax :: Lens' ContactDetail (Maybe Text)

Fax number of the contact.

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code]". For example, a US phone number might appear as "+1.1234567890".

contactDetail_city :: Lens' ContactDetail (Maybe Text)

The city of the contact's address.

contactDetail_state :: Lens' ContactDetail (Maybe Text)

The state or province of the contact's city.

contactDetail_email :: Lens' ContactDetail (Maybe Text)

Email address of the contact.

contactDetail_lastName :: Lens' ContactDetail (Maybe Text)

Last name of contact.

contactDetail_firstName :: Lens' ContactDetail (Maybe Text)

First name of contact.

contactDetail_countryCode :: Lens' ContactDetail (Maybe CountryCode)

Code for the country of the contact's address.

contactDetail_addressLine2 :: Lens' ContactDetail (Maybe Text)

Second line of contact's address, if any.

DomainSuggestion

domainSuggestion_availability :: Lens' DomainSuggestion (Maybe Text)

Whether the domain name is available for registering.

You can register only the domains that are designated as AVAILABLE.

Valid values:

AVAILABLE

The domain name is available.

AVAILABLE_RESERVED

The domain name is reserved under specific conditions.

AVAILABLE_PREORDER

The domain name is available and can be preordered.

DONT_KNOW

The TLD registry didn't reply with a definitive answer about whether the domain name is available. Route 53 can return this response for a variety of reasons, for example, the registry is performing maintenance. Try again later.

PENDING

The TLD registry didn't return a response in the expected amount of time. When the response is delayed, it usually takes just a few extra seconds. You can resubmit the request immediately.

RESERVED

The domain name has been reserved for another person or organization.

UNAVAILABLE

The domain name is not available.

UNAVAILABLE_PREMIUM

The domain name is not available.

UNAVAILABLE_RESTRICTED

The domain name is forbidden.

domainSuggestion_domainName :: Lens' DomainSuggestion (Maybe Text)

A suggested domain name.

DomainSummary

domainSummary_expiry :: Lens' DomainSummary (Maybe UTCTime)

Expiration date of the domain in Unix time format and Coordinated Universal Time (UTC).

domainSummary_autoRenew :: Lens' DomainSummary (Maybe Bool)

Indicates whether the domain is automatically renewed upon expiration.

domainSummary_transferLock :: Lens' DomainSummary (Maybe Bool)

Indicates whether a domain is locked from unauthorized transfer to another party.

domainSummary_domainName :: Lens' DomainSummary Text

The name of the domain that the summary information applies to.

DomainTransferability

domainTransferability_transferable :: Lens' DomainTransferability (Maybe Transferable)

Undocumented member.

ExtraParam

extraParam_name :: Lens' ExtraParam ExtraParamName

The name of an additional parameter that is required by a top-level domain. Here are the top-level domains that require additional parameters and the names of the parameters that they require:

.com.au and .net.au

- AU_ID_NUMBER

AU_ID_TYPE

Valid values include the following:

ABN (Australian business number)

ACN (Australian company number)

TM (Trademark number)

.ca

- BRAND_NUMBER

CA_BUSINESS_ENTITY_TYPE

Valid values include the following:

BANK (Bank)

COMMERCIAL_COMPANY (Commercial company)

COMPANY (Company)

COOPERATION (Cooperation)

COOPERATIVE (Cooperative)

COOPRIX (Cooprix)

CORP (Corporation)

CREDIT_UNION (Credit union)

FOMIA (Federation of mutual insurance associations)

INC (Incorporated)

LTD (Limited)

LTEE (Limitée)

LLC (Limited liability corporation)

LLP (Limited liability partnership)

LTE (Lte.)

MBA (Mutual benefit association)

MIC (Mutual insurance company)

NFP (Not-for-profit corporation)

SA (S.A.)

SAVINGS_COMPANY (Savings company)

SAVINGS_UNION (Savings union)

SARL (Société à responsabilité limitée)

TRUST (Trust)

ULC (Unlimited liability corporation)

CA_LEGAL_TYPE

When ContactType is PERSON, valid values include the following:

ABO (Aboriginal Peoples indigenous to Canada)

CCT (Canadian citizen)

LGR (Legal Representative of a Canadian Citizen or Permanent Resident)

RES (Permanent resident of Canada)

When ContactType is a value other than PERSON, valid values include the following:

ASS (Canadian unincorporated association)

CCO (Canadian corporation)

EDU (Canadian educational institution)

GOV (Government or government entity in Canada)

HOP (Canadian Hospital)

INB (Indian Band recognized by the Indian Act of Canada)

LAM (Canadian Library, Archive, or Museum)

MAJ (Her/His Majesty the Queen/King)

OMK (Official mark registered in Canada)

PLT (Canadian Political Party)

PRT (Partnership Registered in Canada)

TDM (Trademark registered in Canada)

TRD (Canadian Trade Union)

TRS (Trust established in Canada)

.es

- ES_IDENTIFICATION
Specify the applicable value:

For contacts inside Spain: Enter your passport ID.

For contacts outside of Spain: Enter the VAT identification number for the company.

For .es domains, the value of ContactType must be PERSON.

ES_IDENTIFICATION_TYPE

Valid values include the following:

DNI_AND_NIF (For Spanish contacts)

NIE (For foreigners with legal residence)

OTHER (For contacts outside of Spain)

ES_LEGAL_FORM

Valid values include the following:

ASSOCIATION

CENTRAL_GOVERNMENT_BODY

CIVIL_SOCIETY

COMMUNITY_OF_OWNERS

COMMUNITY_PROPERTY

CONSULATE

COOPERATIVE

DESIGNATION_OF_ORIGIN_SUPERVISORY_COUNCIL

ECONOMIC_INTEREST_GROUP

EMBASSY

ENTITY_MANAGING_NATURAL_AREAS

FARM_PARTNERSHIP

FOUNDATION

GENERAL_AND_LIMITED_PARTNERSHIP

GENERAL_PARTNERSHIP

INDIVIDUAL

LIMITED_COMPANY

LOCAL_AUTHORITY

LOCAL_PUBLIC_ENTITY

MUTUAL_INSURANCE_COMPANY

NATIONAL_PUBLIC_ENTITY

ORDER_OR_RELIGIOUS_INSTITUTION

OTHERS (Only for contacts outside of Spain)

POLITICAL_PARTY

PROFESSIONAL_ASSOCIATION

PUBLIC_LAW_ASSOCIATION

PUBLIC_LIMITED_COMPANY

REGIONAL_GOVERNMENT_BODY

REGIONAL_PUBLIC_ENTITY

SAVINGS_BANK

SPANISH_OFFICE

SPORTS_ASSOCIATION

SPORTS_FEDERATION

SPORTS_LIMITED_COMPANY

TEMPORARY_ALLIANCE_OF_ENTERPRISES

TRADE_UNION

WORKER_OWNED_COMPANY

WORKER_OWNED_LIMITED_COMPANY

.fi

- BIRTH_DATE_IN_YYYY_MM_DD

FI_BUSINESS_NUMBER

FI_ID_NUMBER

FI_NATIONALITY

Valid values include the following:

FINNISH

NOT_FINNISH

FI_ORGANIZATION_TYPE

Valid values include the following:

COMPANY

CORPORATION

GOVERNMENT

INSTITUTION

POLITICAL_PARTY

PUBLIC_COMMUNITY

TOWNSHIP

.fr

- BIRTH_CITY

BIRTH_COUNTRY

BIRTH_DATE_IN_YYYY_MM_DD

BIRTH_DEPARTMENT: Specify the INSEE code that corresponds with the department where the contact was born. If the contact was born somewhere other than France or its overseas departments, specify 99. For more information, including a list of departments and the corresponding INSEE numbers, see the Wikipedia entry Departments of France.

BRAND_NUMBER

.it

- IT_NATIONALITY

IT_PIN

IT_REGISTRANT_ENTITY_TYPE

Valid values include the following:

FOREIGNERS

FREELANCE_WORKERS (Freelance workers and professionals)

ITALIAN_COMPANIES (Italian companies and one-person companies)

NON_PROFIT_ORGANIZATIONS

OTHER_SUBJECTS

PUBLIC_ORGANIZATIONS

.ru

- BIRTH_DATE_IN_YYYY_MM_DD

RU_PASSPORT_DATA

.se

- BIRTH_COUNTRY

SE_ID_NUMBER

.sg

- SG_ID_NUMBER

.co.uk, .me.uk, and .org.uk

- UK_CONTACT_TYPE
Valid values include the following:

CRC (UK Corporation by Royal Charter)

FCORP (Non-UK Corporation)

FIND (Non-UK Individual, representing self)

FOTHER (Non-UK Entity that does not fit into any other category)

GOV (UK Government Body)

IND (UK Individual (representing self))

IP (UK Industrial/Provident Registered Company)

LLP (UK Limited Liability Partnership)

LTD (UK Limited Company)

OTHER (UK Entity that does not fit into any other category)

PLC (UK Public Limited Company)

PTNR (UK Partnership)

RCHAR (UK Registered Charity)

SCH (UK School)

STAT (UK Statutory Body)

STRA (UK Sole Trader)

UK_COMPANY_NUMBER

In addition, many TLDs require a VAT_NUMBER.

extraParam_value :: Lens' ExtraParam Text

The value that corresponds with the name of an extra parameter.

Nameserver

nameserver_glueIps :: Lens' Nameserver (Maybe [Text])

Glue IP address of a name server entry. Glue IP addresses are required only when the name of the name server is a subdomain of the domain. For example, if your domain is example.com and the name server for the domain is ns.example.com, you need to specify the IP address for ns.example.com.

Constraints: The list can contain only one IPv4 and one IPv6 address.

nameserver_name :: Lens' Nameserver Text

The fully qualified host name of the name server.

Constraint: Maximum 255 characters

OperationSummary

operationSummary_operationId :: Lens' OperationSummary Text

Identifier returned to track the requested action.

operationSummary_status :: Lens' OperationSummary OperationStatus

The current status of the requested operation in the system.

operationSummary_type :: Lens' OperationSummary OperationType

Type of the action requested.

operationSummary_submittedDate :: Lens' OperationSummary UTCTime

The date when the request was submitted.

Tag

tag_key :: Lens' Tag (Maybe Text)

The key (name) of a tag.

Valid values: A-Z, a-z, 0-9, space, ".:/=+\-@"

Constraints: Each key can be 1-128 characters long.

tag_value :: Lens' Tag (Maybe Text)

The value of a tag.

Valid values: A-Z, a-z, 0-9, space, ".:/=+\-@"

Constraints: Each value can be 0-256 characters long.