Network.AWS.Route53Domains.RegisterDomain

Description

This operation registers a domain. Domains are registered either by Amazon Registrar (for .com, .net, and .org domains) or by our registrar associate, Gandi (for all other domains). For some top-level domains (TLDs), this operation requires extra parameters.

When you register a domain, Amazon Route 53 does the following:

Creates a Route 53 hosted zone that has the same name as the domain. Route 53 assigns four name servers to your hosted zone and automatically updates your domain registration with the names of these name servers.

Enables autorenew, so your domain registration will renew automatically each year. We'll notify you in advance of the renewal date so you can choose whether to renew the registration.

Optionally enables privacy protection, so WHOIS queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you don't enable privacy protection, WHOIS queries return the information that you entered for the registrant, admin, and tech contacts.

If registration is successful, returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant is notified by email.

Charges your AWS account an amount based on the top-level domain. For more information, see Amazon Route 53 Pricing.

Creating a Request

data RegisterDomain

The RegisterDomain request includes the following elements.

See: newRegisterDomain smart constructor.

Constructors

RegisterDomain'

Fields

privacyProtectTechContact :: Maybe Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the technical contact.

Default: true

privacyProtectRegistrantContact :: Maybe Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the registrant contact (the domain owner).

Default: true

autoRenew :: Maybe Bool

Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged.

Default: true

privacyProtectAdminContact :: Maybe Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the admin contact.

Default: true

idnLangCode :: Maybe Text

Reserved for future use.

domainName :: Text

The domain name that you want to register. The top-level domain (TLD), such as .com, must be a TLD that Route 53 supports. For a list of supported TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

The domain name can contain only the following characters:

Letters a through z. Domain names are not case sensitive.

Numbers 0 through 9.

Hyphen (-). You can't specify a hyphen at the beginning or end of a label.

Period (.) to separate the labels in the name, such as the . in example.com.

Internationalized domain names are not supported for some top-level domains. To determine whether the TLD that you want to use supports internationalized domain names, see Domains that You Can Register with Amazon Route 53. For more information, see Formatting Internationalized Domain Names.

durationInYears :: Natural

The number of years that you want to register the domain for. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain. For the range of valid values for your domain, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

Default: 1

adminContact :: Sensitive ContactDetail

Provides detailed contact information. For information about the values that you specify for each element, see ContactDetail.

registrantContact :: Sensitive ContactDetail

Provides detailed contact information. For information about the values that you specify for each element, see ContactDetail.

techContact :: Sensitive ContactDetail

Provides detailed contact information. For information about the values that you specify for each element, see ContactDetail.

newRegisterDomain

Arguments

:: Text
$sel:domainName:RegisterDomain'

-> Natural
$sel:durationInYears:RegisterDomain'

-> ContactDetail
$sel:adminContact:RegisterDomain'

-> ContactDetail
$sel:registrantContact:RegisterDomain'

-> ContactDetail
$sel:techContact:RegisterDomain'

-> RegisterDomain

Create a value of RegisterDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:privacyProtectTechContact:RegisterDomain', registerDomain_privacyProtectTechContact - Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the technical contact.

Default: true

$sel:privacyProtectRegistrantContact:RegisterDomain', registerDomain_privacyProtectRegistrantContact - Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the registrant contact (the domain owner).

Default: true

$sel:autoRenew:RegisterDomain', registerDomain_autoRenew - Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged.

Default: true

$sel:privacyProtectAdminContact:RegisterDomain', registerDomain_privacyProtectAdminContact - Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the admin contact.

Default: true

$sel:idnLangCode:RegisterDomain', registerDomain_idnLangCode - Reserved for future use.

$sel:domainName:RegisterDomain', registerDomain_domainName - The domain name that you want to register. The top-level domain (TLD), such as .com, must be a TLD that Route 53 supports. For a list of supported TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

The domain name can contain only the following characters:

Letters a through z. Domain names are not case sensitive.

Numbers 0 through 9.

Hyphen (-). You can't specify a hyphen at the beginning or end of a label.

Period (.) to separate the labels in the name, such as the . in example.com.

Internationalized domain names are not supported for some top-level domains. To determine whether the TLD that you want to use supports internationalized domain names, see Domains that You Can Register with Amazon Route 53. For more information, see Formatting Internationalized Domain Names.

$sel:durationInYears:RegisterDomain', registerDomain_durationInYears - The number of years that you want to register the domain for. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain. For the range of valid values for your domain, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

Default: 1

$sel:adminContact:RegisterDomain', registerDomain_adminContact - Provides detailed contact information. For information about the values that you specify for each element, see ContactDetail.

$sel:registrantContact:RegisterDomain', registerDomain_registrantContact - Provides detailed contact information. For information about the values that you specify for each element, see ContactDetail.

$sel:techContact:RegisterDomain', registerDomain_techContact - Provides detailed contact information. For information about the values that you specify for each element, see ContactDetail.

Request Lenses

registerDomain_privacyProtectTechContact :: Lens' RegisterDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the technical contact.

Default: true

registerDomain_privacyProtectRegistrantContact :: Lens' RegisterDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the registrant contact (the domain owner).

Default: true

registerDomain_autoRenew :: Lens' RegisterDomain (Maybe Bool)

Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged.

Default: true

registerDomain_privacyProtectAdminContact :: Lens' RegisterDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the admin contact.

Default: true

registerDomain_idnLangCode :: Lens' RegisterDomain (Maybe Text)

Reserved for future use.

registerDomain_domainName :: Lens' RegisterDomain Text

The domain name that you want to register. The top-level domain (TLD), such as .com, must be a TLD that Route 53 supports. For a list of supported TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

The domain name can contain only the following characters:

Letters a through z. Domain names are not case sensitive.

Numbers 0 through 9.

Hyphen (-). You can't specify a hyphen at the beginning or end of a label.

Period (.) to separate the labels in the name, such as the . in example.com.

Internationalized domain names are not supported for some top-level domains. To determine whether the TLD that you want to use supports internationalized domain names, see Domains that You Can Register with Amazon Route 53. For more information, see Formatting Internationalized Domain Names.

registerDomain_durationInYears :: Lens' RegisterDomain Natural

The number of years that you want to register the domain for. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain. For the range of valid values for your domain, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

Default: 1

registerDomain_adminContact :: Lens' RegisterDomain ContactDetail

Provides detailed contact information. For information about the values that you specify for each element, see ContactDetail.

registerDomain_registrantContact :: Lens' RegisterDomain ContactDetail

Provides detailed contact information. For information about the values that you specify for each element, see ContactDetail.

registerDomain_techContact :: Lens' RegisterDomain ContactDetail

Provides detailed contact information. For information about the values that you specify for each element, see ContactDetail.

Destructuring the Response

data RegisterDomainResponse

The RegisterDomain response includes the following element.

See: newRegisterDomainResponse smart constructor.

Constructors

RegisterDomainResponse'

Fields

httpStatus :: Int

The response's http status code.

operationId :: Text

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

newRegisterDomainResponse

Arguments

:: Int
$sel:httpStatus:RegisterDomainResponse'

-> Text
$sel:operationId:RegisterDomainResponse'

-> RegisterDomainResponse

Create a value of RegisterDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RegisterDomainResponse', registerDomainResponse_httpStatus - The response's http status code.

$sel:operationId:RegisterDomainResponse', registerDomainResponse_operationId - Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

Response Lenses

registerDomainResponse_httpStatus :: Lens' RegisterDomainResponse Int

The response's http status code.

registerDomainResponse_operationId :: Lens' RegisterDomainResponse Text

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.