Network.AWS.Route53Domains.RenewDomain

Description

This operation renews a domain for the specified number of years. The cost of renewing your domain is billed to your AWS account.

We recommend that you renew your domain several weeks before the expiration date. Some TLD registries delete domains before the expiration date if you haven't renewed far enough in advance. For more information about renewing domain registration, see Renewing Registration for a Domain in the Amazon Route 53 Developer Guide.

Creating a Request

data RenewDomain

A RenewDomain request includes the number of years that you want to renew for and the current expiration year.

See: newRenewDomain smart constructor.

Constructors

RenewDomain'

Fields

durationInYears :: Maybe Natural

The number of years that you want to renew the domain for. The maximum number of years depends on the top-level domain. For the range of valid values for your domain, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

Default: 1

domainName :: Text

The name of the domain that you want to renew.

currentExpiryYear :: Int

The year when the registration for the domain is set to expire. This value must match the current expiration date for the domain.

newRenewDomain

Arguments

:: Text
$sel:domainName:RenewDomain'

-> Int
$sel:currentExpiryYear:RenewDomain'

-> RenewDomain

Create a value of RenewDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:durationInYears:RenewDomain', renewDomain_durationInYears - The number of years that you want to renew the domain for. The maximum number of years depends on the top-level domain. For the range of valid values for your domain, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

Default: 1

$sel:domainName:RenewDomain', renewDomain_domainName - The name of the domain that you want to renew.

$sel:currentExpiryYear:RenewDomain', renewDomain_currentExpiryYear - The year when the registration for the domain is set to expire. This value must match the current expiration date for the domain.

Request Lenses

renewDomain_durationInYears :: Lens' RenewDomain (Maybe Natural)

The number of years that you want to renew the domain for. The maximum number of years depends on the top-level domain. For the range of valid values for your domain, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

Default: 1

renewDomain_domainName :: Lens' RenewDomain Text

The name of the domain that you want to renew.

renewDomain_currentExpiryYear :: Lens' RenewDomain Int

The year when the registration for the domain is set to expire. This value must match the current expiration date for the domain.

Destructuring the Response

data RenewDomainResponse

See: newRenewDomainResponse smart constructor.

Constructors

RenewDomainResponse'

Fields

httpStatus :: Int

The response's http status code.

operationId :: Text

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

newRenewDomainResponse

Arguments

:: Int
$sel:httpStatus:RenewDomainResponse'

-> Text
$sel:operationId:RenewDomainResponse'

-> RenewDomainResponse

Create a value of RenewDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RenewDomainResponse', renewDomainResponse_httpStatus - The response's http status code.

$sel:operationId:RenewDomainResponse', renewDomainResponse_operationId - Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

Response Lenses

renewDomainResponse_httpStatus :: Lens' RenewDomainResponse Int

The response's http status code.

renewDomainResponse_operationId :: Lens' RenewDomainResponse Text

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.