Network.AWS.Route53Domains.ResendContactReachabilityEmail

Description

For operations that require confirmation that the email address for the registrant contact is valid, such as registering a new domain, this operation resends the confirmation email to the current email address for the registrant contact.

Creating a Request

data ResendContactReachabilityEmail

See: newResendContactReachabilityEmail smart constructor.

Constructors

ResendContactReachabilityEmail'

Fields

domainName :: Maybe Text

The name of the domain for which you want Route 53 to resend a confirmation email to the registrant contact.

newResendContactReachabilityEmail :: ResendContactReachabilityEmail

Create a value of ResendContactReachabilityEmail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:ResendContactReachabilityEmail', resendContactReachabilityEmail_domainName - The name of the domain for which you want Route 53 to resend a confirmation email to the registrant contact.

Request Lenses

resendContactReachabilityEmail_domainName :: Lens' ResendContactReachabilityEmail (Maybe Text)

The name of the domain for which you want Route 53 to resend a confirmation email to the registrant contact.

Destructuring the Response

data ResendContactReachabilityEmailResponse

See: newResendContactReachabilityEmailResponse smart constructor.

Constructors

ResendContactReachabilityEmailResponse'

Fields

isAlreadyVerified :: Maybe Bool

True if the email address for the registrant contact has already been verified, and false otherwise. If the email address has already been verified, we don't send another confirmation email.

domainName :: Maybe Text

The domain name for which you requested a confirmation email.

emailAddress :: Maybe Text

The email address for the registrant contact at the time that we sent the verification email.

httpStatus :: Int

The response's http status code.

newResendContactReachabilityEmailResponse

Arguments

:: Int
$sel:httpStatus:ResendContactReachabilityEmailResponse'

-> ResendContactReachabilityEmailResponse

Create a value of ResendContactReachabilityEmailResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isAlreadyVerified:ResendContactReachabilityEmailResponse', resendContactReachabilityEmailResponse_isAlreadyVerified - True if the email address for the registrant contact has already been verified, and false otherwise. If the email address has already been verified, we don't send another confirmation email.

$sel:domainName:ResendContactReachabilityEmail', resendContactReachabilityEmailResponse_domainName - The domain name for which you requested a confirmation email.

$sel:emailAddress:ResendContactReachabilityEmailResponse', resendContactReachabilityEmailResponse_emailAddress - The email address for the registrant contact at the time that we sent the verification email.

$sel:httpStatus:ResendContactReachabilityEmailResponse', resendContactReachabilityEmailResponse_httpStatus - The response's http status code.

Response Lenses

resendContactReachabilityEmailResponse_isAlreadyVerified :: Lens' ResendContactReachabilityEmailResponse (Maybe Bool)

True if the email address for the registrant contact has already been verified, and false otherwise. If the email address has already been verified, we don't send another confirmation email.

resendContactReachabilityEmailResponse_domainName :: Lens' ResendContactReachabilityEmailResponse (Maybe Text)

The domain name for which you requested a confirmation email.

resendContactReachabilityEmailResponse_emailAddress :: Lens' ResendContactReachabilityEmailResponse (Maybe Text)

The email address for the registrant contact at the time that we sent the verification email.

resendContactReachabilityEmailResponse_httpStatus :: Lens' ResendContactReachabilityEmailResponse Int

The response's http status code.