Network.AWS.Route53Domains.RetrieveDomainAuthCode

Description

This operation returns the AuthCode for the domain. To transfer a domain to another registrar, you provide this value to the new registrar.

Creating a Request

data RetrieveDomainAuthCode

A request for the authorization code for the specified domain. To transfer a domain to another registrar, you provide this value to the new registrar.

See: newRetrieveDomainAuthCode smart constructor.

Constructors

RetrieveDomainAuthCode'

Fields

domainName :: Text

The name of the domain that you want to get an authorization code for.

newRetrieveDomainAuthCode

Arguments

:: Text
$sel:domainName:RetrieveDomainAuthCode'

-> RetrieveDomainAuthCode

Create a value of RetrieveDomainAuthCode with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:RetrieveDomainAuthCode', retrieveDomainAuthCode_domainName - The name of the domain that you want to get an authorization code for.

Request Lenses

retrieveDomainAuthCode_domainName :: Lens' RetrieveDomainAuthCode Text

The name of the domain that you want to get an authorization code for.

Destructuring the Response

data RetrieveDomainAuthCodeResponse

The RetrieveDomainAuthCode response includes the following element.

See: newRetrieveDomainAuthCodeResponse smart constructor.

Constructors

RetrieveDomainAuthCodeResponse'

Fields

httpStatus :: Int

The response's http status code.

authCode :: Sensitive Text

The authorization code for the domain.

newRetrieveDomainAuthCodeResponse

Arguments

:: Int
$sel:httpStatus:RetrieveDomainAuthCodeResponse'

-> Text
$sel:authCode:RetrieveDomainAuthCodeResponse'

-> RetrieveDomainAuthCodeResponse

Create a value of RetrieveDomainAuthCodeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RetrieveDomainAuthCodeResponse', retrieveDomainAuthCodeResponse_httpStatus - The response's http status code.

$sel:authCode:RetrieveDomainAuthCodeResponse', retrieveDomainAuthCodeResponse_authCode - The authorization code for the domain.

Response Lenses

retrieveDomainAuthCodeResponse_httpStatus :: Lens' RetrieveDomainAuthCodeResponse Int

The response's http status code.

retrieveDomainAuthCodeResponse_authCode :: Lens' RetrieveDomainAuthCodeResponse Text

The authorization code for the domain.