Network.AWS.Route53Domains.Types

Description

Service Configuration

defaultService :: Service

API version 2014-05-15 of the Amazon Route 53 Domains SDK configuration.

Errors

_InvalidInput :: AsError a => Getting (First ServiceError) a ServiceError

The requested item is not acceptable. For example, for APIs that accept a domain name, the request might specify a domain name that doesn't belong to the account that submitted the request. For AcceptDomainTransferFromAnotherAwsAccount, the password might be invalid.

_UnsupportedTLD :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Route 53 does not support this top-level domain (TLD).

_DuplicateRequest :: AsError a => Getting (First ServiceError) a ServiceError

The request is already in progress for the domain.

_DomainLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

The number of domains has exceeded the allowed threshold for the account.

_OperationLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

The number of operations or jobs running exceeded the allowed threshold for the account.

_TLDRulesViolation :: AsError a => Getting (First ServiceError) a ServiceError

The top-level domain does not support this operation.

ContactType

newtype ContactType

Constructors

ContactType'

Fields

fromContactType :: Text

Bundled Patterns

pattern ContactType_ASSOCIATION :: ContactType

pattern ContactType_COMPANY :: ContactType

pattern ContactType_PERSON :: ContactType

pattern ContactType_PUBLIC_BODY :: ContactType

pattern ContactType_RESELLER :: ContactType

CountryCode

newtype CountryCode

Constructors

CountryCode'

Fields

fromCountryCode :: Text

Bundled Patterns

pattern CountryCode_AD :: CountryCode

pattern CountryCode_AE :: CountryCode

pattern CountryCode_AF :: CountryCode

pattern CountryCode_AG :: CountryCode

pattern CountryCode_AI :: CountryCode

pattern CountryCode_AL :: CountryCode

pattern CountryCode_AM :: CountryCode

pattern CountryCode_AN :: CountryCode

pattern CountryCode_AO :: CountryCode

pattern CountryCode_AQ :: CountryCode

pattern CountryCode_AR :: CountryCode

pattern CountryCode_AS :: CountryCode

pattern CountryCode_AT :: CountryCode

pattern CountryCode_AU :: CountryCode

pattern CountryCode_AW :: CountryCode

pattern CountryCode_AZ :: CountryCode

pattern CountryCode_BA :: CountryCode

pattern CountryCode_BB :: CountryCode

pattern CountryCode_BD :: CountryCode

pattern CountryCode_BE :: CountryCode

pattern CountryCode_BF :: CountryCode

pattern CountryCode_BG :: CountryCode

pattern CountryCode_BH :: CountryCode

pattern CountryCode_BI :: CountryCode

pattern CountryCode_BJ :: CountryCode

pattern CountryCode_BL :: CountryCode

pattern CountryCode_BM :: CountryCode

pattern CountryCode_BN :: CountryCode

pattern CountryCode_BO :: CountryCode

pattern CountryCode_BR :: CountryCode

pattern CountryCode_BS :: CountryCode

pattern CountryCode_BT :: CountryCode

pattern CountryCode_BW :: CountryCode

pattern CountryCode_BY :: CountryCode

pattern CountryCode_BZ :: CountryCode

pattern CountryCode_CA :: CountryCode

pattern CountryCode_CC :: CountryCode

pattern CountryCode_CD :: CountryCode

pattern CountryCode_CF :: CountryCode

pattern CountryCode_CG :: CountryCode

pattern CountryCode_CH :: CountryCode

pattern CountryCode_CI :: CountryCode

pattern CountryCode_CK :: CountryCode

pattern CountryCode_CL :: CountryCode

pattern CountryCode_CM :: CountryCode

pattern CountryCode_CN :: CountryCode

pattern CountryCode_CO :: CountryCode

pattern CountryCode_CR :: CountryCode

pattern CountryCode_CU :: CountryCode

pattern CountryCode_CV :: CountryCode

pattern CountryCode_CX :: CountryCode

pattern CountryCode_CY :: CountryCode

pattern CountryCode_CZ :: CountryCode

pattern CountryCode_DE :: CountryCode

pattern CountryCode_DJ :: CountryCode

pattern CountryCode_DK :: CountryCode

pattern CountryCode_DM :: CountryCode

pattern CountryCode_DO :: CountryCode

pattern CountryCode_DZ :: CountryCode

pattern CountryCode_EC :: CountryCode

pattern CountryCode_EE :: CountryCode

pattern CountryCode_EG :: CountryCode

pattern CountryCode_ER :: CountryCode

pattern CountryCode_ES :: CountryCode

pattern CountryCode_ET :: CountryCode

pattern CountryCode_FI :: CountryCode

pattern CountryCode_FJ :: CountryCode

pattern CountryCode_FK :: CountryCode

pattern CountryCode_FM :: CountryCode

pattern CountryCode_FO :: CountryCode

pattern CountryCode_FR :: CountryCode

pattern CountryCode_GA :: CountryCode

pattern CountryCode_GB :: CountryCode

pattern CountryCode_GD :: CountryCode

pattern CountryCode_GE :: CountryCode

pattern CountryCode_GH :: CountryCode

pattern CountryCode_GI :: CountryCode

pattern CountryCode_GL :: CountryCode

pattern CountryCode_GM :: CountryCode

pattern CountryCode_GN :: CountryCode

pattern CountryCode_GQ :: CountryCode

pattern CountryCode_GR :: CountryCode

pattern CountryCode_GT :: CountryCode

pattern CountryCode_GU :: CountryCode

pattern CountryCode_GW :: CountryCode

pattern CountryCode_GY :: CountryCode

pattern CountryCode_HK :: CountryCode

pattern CountryCode_HN :: CountryCode

pattern CountryCode_HR :: CountryCode

pattern CountryCode_HT :: CountryCode

pattern CountryCode_HU :: CountryCode

pattern CountryCode_ID :: CountryCode

pattern CountryCode_IE :: CountryCode

pattern CountryCode_IL :: CountryCode

pattern CountryCode_IM :: CountryCode

pattern CountryCode_IN :: CountryCode

pattern CountryCode_IQ :: CountryCode

pattern CountryCode_IR :: CountryCode

pattern CountryCode_IS :: CountryCode

pattern CountryCode_IT :: CountryCode

pattern CountryCode_JM :: CountryCode

pattern CountryCode_JO :: CountryCode

pattern CountryCode_JP :: CountryCode

pattern CountryCode_KE :: CountryCode

pattern CountryCode_KG :: CountryCode

pattern CountryCode_KH :: CountryCode

pattern CountryCode_KI :: CountryCode

pattern CountryCode_KM :: CountryCode

pattern CountryCode_KN :: CountryCode

pattern CountryCode_KP :: CountryCode

pattern CountryCode_KR :: CountryCode

pattern CountryCode_KW :: CountryCode

pattern CountryCode_KY :: CountryCode

pattern CountryCode_KZ :: CountryCode

pattern CountryCode_LA :: CountryCode

pattern CountryCode_LB :: CountryCode

pattern CountryCode_LC :: CountryCode

pattern CountryCode_LI :: CountryCode

pattern CountryCode_LK :: CountryCode

pattern CountryCode_LR :: CountryCode

pattern CountryCode_LS :: CountryCode

pattern CountryCode_LT :: CountryCode

pattern CountryCode_LU :: CountryCode

pattern CountryCode_LV :: CountryCode

pattern CountryCode_LY :: CountryCode

pattern CountryCode_MA :: CountryCode

pattern CountryCode_MC :: CountryCode

pattern CountryCode_MD :: CountryCode

pattern CountryCode_ME :: CountryCode

pattern CountryCode_MF :: CountryCode

pattern CountryCode_MG :: CountryCode

pattern CountryCode_MH :: CountryCode

pattern CountryCode_MK :: CountryCode

pattern CountryCode_ML :: CountryCode

pattern CountryCode_MM :: CountryCode

pattern CountryCode_MN :: CountryCode

pattern CountryCode_MO :: CountryCode

pattern CountryCode_MP :: CountryCode

pattern CountryCode_MR :: CountryCode

pattern CountryCode_MS :: CountryCode

pattern CountryCode_MT :: CountryCode

pattern CountryCode_MU :: CountryCode

pattern CountryCode_MV :: CountryCode

pattern CountryCode_MW :: CountryCode

pattern CountryCode_MX :: CountryCode

pattern CountryCode_MY :: CountryCode

pattern CountryCode_MZ :: CountryCode

pattern CountryCode_NA :: CountryCode

pattern CountryCode_NC :: CountryCode

pattern CountryCode_NE :: CountryCode

pattern CountryCode_NG :: CountryCode

pattern CountryCode_NI :: CountryCode

pattern CountryCode_NL :: CountryCode

pattern CountryCode_NO :: CountryCode

pattern CountryCode_NP :: CountryCode

pattern CountryCode_NR :: CountryCode

pattern CountryCode_NU :: CountryCode

pattern CountryCode_NZ :: CountryCode

pattern CountryCode_OM :: CountryCode

pattern CountryCode_PA :: CountryCode

pattern CountryCode_PE :: CountryCode

pattern CountryCode_PF :: CountryCode

pattern CountryCode_PG :: CountryCode

pattern CountryCode_PH :: CountryCode

pattern CountryCode_PK :: CountryCode

pattern CountryCode_PL :: CountryCode

pattern CountryCode_PM :: CountryCode

pattern CountryCode_PN :: CountryCode

pattern CountryCode_PR :: CountryCode

pattern CountryCode_PT :: CountryCode

pattern CountryCode_PW :: CountryCode

pattern CountryCode_PY :: CountryCode

pattern CountryCode_QA :: CountryCode

pattern CountryCode_RO :: CountryCode

pattern CountryCode_RS :: CountryCode

pattern CountryCode_RU :: CountryCode

pattern CountryCode_RW :: CountryCode

pattern CountryCode_SA :: CountryCode

pattern CountryCode_SB :: CountryCode

pattern CountryCode_SC :: CountryCode

pattern CountryCode_SD :: CountryCode

pattern CountryCode_SE :: CountryCode

pattern CountryCode_SG :: CountryCode

pattern CountryCode_SH :: CountryCode

pattern CountryCode_SI :: CountryCode

pattern CountryCode_SK :: CountryCode

pattern CountryCode_SL :: CountryCode

pattern CountryCode_SM :: CountryCode

pattern CountryCode_SN :: CountryCode

pattern CountryCode_SO :: CountryCode

pattern CountryCode_SR :: CountryCode

pattern CountryCode_ST :: CountryCode

pattern CountryCode_SV :: CountryCode

pattern CountryCode_SY :: CountryCode

pattern CountryCode_SZ :: CountryCode

pattern CountryCode_TC :: CountryCode

pattern CountryCode_TD :: CountryCode

pattern CountryCode_TG :: CountryCode

pattern CountryCode_TH :: CountryCode

pattern CountryCode_TJ :: CountryCode

pattern CountryCode_TK :: CountryCode

pattern CountryCode_TL :: CountryCode

pattern CountryCode_TM :: CountryCode

pattern CountryCode_TN :: CountryCode

pattern CountryCode_TO :: CountryCode

pattern CountryCode_TR :: CountryCode

pattern CountryCode_TT :: CountryCode

pattern CountryCode_TV :: CountryCode

pattern CountryCode_TW :: CountryCode

pattern CountryCode_TZ :: CountryCode

pattern CountryCode_UA :: CountryCode

pattern CountryCode_UG :: CountryCode

pattern CountryCode_US :: CountryCode

pattern CountryCode_UY :: CountryCode

pattern CountryCode_UZ :: CountryCode

pattern CountryCode_VA :: CountryCode

pattern CountryCode_VC :: CountryCode

pattern CountryCode_VE :: CountryCode

pattern CountryCode_VG :: CountryCode

pattern CountryCode_VI :: CountryCode

pattern CountryCode_VN :: CountryCode

pattern CountryCode_VU :: CountryCode

pattern CountryCode_WF :: CountryCode

pattern CountryCode_WS :: CountryCode

pattern CountryCode_YE :: CountryCode

pattern CountryCode_YT :: CountryCode

pattern CountryCode_ZA :: CountryCode

pattern CountryCode_ZM :: CountryCode

pattern CountryCode_ZW :: CountryCode

DomainAvailability

newtype DomainAvailability

Constructors

DomainAvailability'

Fields

fromDomainAvailability :: Text

Bundled Patterns

pattern DomainAvailability_AVAILABLE :: DomainAvailability

pattern DomainAvailability_AVAILABLE_PREORDER :: DomainAvailability

pattern DomainAvailability_AVAILABLE_RESERVED :: DomainAvailability

pattern DomainAvailability_DONT_KNOW :: DomainAvailability

pattern DomainAvailability_RESERVED :: DomainAvailability

pattern DomainAvailability_UNAVAILABLE :: DomainAvailability

pattern DomainAvailability_UNAVAILABLE_PREMIUM :: DomainAvailability

pattern DomainAvailability_UNAVAILABLE_RESTRICTED :: DomainAvailability

ExtraParamName

newtype ExtraParamName

Constructors

ExtraParamName'

Fields

fromExtraParamName :: Text

Bundled Patterns

pattern ExtraParamName_AU_ID_NUMBER :: ExtraParamName

pattern ExtraParamName_AU_ID_TYPE :: ExtraParamName

pattern ExtraParamName_BIRTH_CITY :: ExtraParamName

pattern ExtraParamName_BIRTH_COUNTRY :: ExtraParamName

pattern ExtraParamName_BIRTH_DATE_IN_YYYY_MM_DD :: ExtraParamName

pattern ExtraParamName_BIRTH_DEPARTMENT :: ExtraParamName

pattern ExtraParamName_BRAND_NUMBER :: ExtraParamName

pattern ExtraParamName_CA_BUSINESS_ENTITY_TYPE :: ExtraParamName

pattern ExtraParamName_CA_LEGAL_REPRESENTATIVE :: ExtraParamName

pattern ExtraParamName_CA_LEGAL_REPRESENTATIVE_CAPACITY :: ExtraParamName

pattern ExtraParamName_CA_LEGAL_TYPE :: ExtraParamName

pattern ExtraParamName_DOCUMENT_NUMBER :: ExtraParamName

pattern ExtraParamName_DUNS_NUMBER :: ExtraParamName

pattern ExtraParamName_ES_IDENTIFICATION :: ExtraParamName

pattern ExtraParamName_ES_IDENTIFICATION_TYPE :: ExtraParamName

pattern ExtraParamName_ES_LEGAL_FORM :: ExtraParamName

pattern ExtraParamName_FI_BUSINESS_NUMBER :: ExtraParamName

pattern ExtraParamName_FI_ID_NUMBER :: ExtraParamName

pattern ExtraParamName_FI_NATIONALITY :: ExtraParamName

pattern ExtraParamName_FI_ORGANIZATION_TYPE :: ExtraParamName

pattern ExtraParamName_IT_NATIONALITY :: ExtraParamName

pattern ExtraParamName_IT_PIN :: ExtraParamName

pattern ExtraParamName_IT_REGISTRANT_ENTITY_TYPE :: ExtraParamName

pattern ExtraParamName_RU_PASSPORT_DATA :: ExtraParamName

pattern ExtraParamName_SE_ID_NUMBER :: ExtraParamName

pattern ExtraParamName_SG_ID_NUMBER :: ExtraParamName

pattern ExtraParamName_UK_COMPANY_NUMBER :: ExtraParamName

pattern ExtraParamName_UK_CONTACT_TYPE :: ExtraParamName

pattern ExtraParamName_VAT_NUMBER :: ExtraParamName

OperationStatus

newtype OperationStatus

Constructors

OperationStatus'

Fields

fromOperationStatus :: Text

Bundled Patterns

pattern OperationStatus_ERROR :: OperationStatus

pattern OperationStatus_FAILED :: OperationStatus

pattern OperationStatus_IN_PROGRESS :: OperationStatus

pattern OperationStatus_SUBMITTED :: OperationStatus

pattern OperationStatus_SUCCESSFUL :: OperationStatus

OperationType

newtype OperationType

Constructors

OperationType'

Fields

fromOperationType :: Text

Bundled Patterns

pattern OperationType_ADD_DNSSEC :: OperationType

pattern OperationType_CHANGE_DOMAIN_OWNER :: OperationType

pattern OperationType_CHANGE_PRIVACY_PROTECTION :: OperationType

pattern OperationType_DELETE_DOMAIN :: OperationType

pattern OperationType_DISABLE_AUTORENEW :: OperationType

pattern OperationType_DOMAIN_LOCK :: OperationType

pattern OperationType_ENABLE_AUTORENEW :: OperationType

pattern OperationType_EXPIRE_DOMAIN :: OperationType

pattern OperationType_INTERNAL_TRANSFER_IN_DOMAIN :: OperationType

pattern OperationType_INTERNAL_TRANSFER_OUT_DOMAIN :: OperationType

pattern OperationType_PUSH_DOMAIN :: OperationType

pattern OperationType_REGISTER_DOMAIN :: OperationType

pattern OperationType_REMOVE_DNSSEC :: OperationType

pattern OperationType_RENEW_DOMAIN :: OperationType

pattern OperationType_TRANSFER_IN_DOMAIN :: OperationType

pattern OperationType_TRANSFER_OUT_DOMAIN :: OperationType

pattern OperationType_UPDATE_DOMAIN_CONTACT :: OperationType

pattern OperationType_UPDATE_NAMESERVER :: OperationType

ReachabilityStatus

newtype ReachabilityStatus

Constructors

ReachabilityStatus'

Fields

fromReachabilityStatus :: Text

Bundled Patterns

pattern ReachabilityStatus_DONE :: ReachabilityStatus

pattern ReachabilityStatus_EXPIRED :: ReachabilityStatus

pattern ReachabilityStatus_PENDING :: ReachabilityStatus

Transferable

newtype Transferable

Whether the domain name can be transferred to Route 53.

You can transfer only domains that have a value of TRANSFERABLE for Transferable.

Valid values:

TRANSFERABLE

The domain name can be transferred to Route 53.

UNTRANSFERRABLE

The domain name can't be transferred to Route 53.

DONT_KNOW

Reserved for future use.

Constructors

Transferable'

Fields

fromTransferable :: Text

Bundled Patterns

pattern Transferable_DONT_KNOW :: Transferable

pattern Transferable_TRANSFERABLE :: Transferable

pattern Transferable_UNTRANSFERABLE :: Transferable

BillingRecord

data BillingRecord

Information for one billing record.

See: newBillingRecord smart constructor.

Constructors

BillingRecord'

Fields

invoiceId :: Maybe Text

The ID of the invoice that is associated with the billing record.

operation :: Maybe OperationType

The operation that you were charged for.

billDate :: Maybe POSIX

The date that the operation was billed, in Unix format.

domainName :: Maybe Text

The name of the domain that the billing record applies to. If the domain name contains characters other than a-z, 0-9, and - (hyphen), such as an internationalized domain name, then this value is in Punycode. For more information, see DNS Domain Name Format in the Amazon Route 53 Developer Guide.

price :: Maybe Double

The price that you were charged for the operation, in US dollars.

Example value: 12.0

newBillingRecord :: BillingRecord

Create a value of BillingRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:invoiceId:BillingRecord', billingRecord_invoiceId - The ID of the invoice that is associated with the billing record.

$sel:operation:BillingRecord', billingRecord_operation - The operation that you were charged for.

$sel:billDate:BillingRecord', billingRecord_billDate - The date that the operation was billed, in Unix format.

$sel:domainName:BillingRecord', billingRecord_domainName - The name of the domain that the billing record applies to. If the domain name contains characters other than a-z, 0-9, and - (hyphen), such as an internationalized domain name, then this value is in Punycode. For more information, see DNS Domain Name Format in the Amazon Route 53 Developer Guide.

$sel:price:BillingRecord', billingRecord_price - The price that you were charged for the operation, in US dollars.

Example value: 12.0

billingRecord_invoiceId :: Lens' BillingRecord (Maybe Text)

The ID of the invoice that is associated with the billing record.

billingRecord_operation :: Lens' BillingRecord (Maybe OperationType)

The operation that you were charged for.

billingRecord_billDate :: Lens' BillingRecord (Maybe UTCTime)

The date that the operation was billed, in Unix format.

billingRecord_domainName :: Lens' BillingRecord (Maybe Text)

The name of the domain that the billing record applies to. If the domain name contains characters other than a-z, 0-9, and - (hyphen), such as an internationalized domain name, then this value is in Punycode. For more information, see DNS Domain Name Format in the Amazon Route 53 Developer Guide.

billingRecord_price :: Lens' BillingRecord (Maybe Double)

The price that you were charged for the operation, in US dollars.

Example value: 12.0

ContactDetail

data ContactDetail

ContactDetail includes the following elements.

See: newContactDetail smart constructor.

Constructors

ContactDetail'

Fields

phoneNumber :: Maybe Text

The phone number of the contact.

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code>]". For example, a US phone number might appear as "+1.1234567890".

organizationName :: Maybe Text

Name of the organization for contact types other than PERSON.

addressLine1 :: Maybe Text

First line of the contact's address.

extraParams :: Maybe [ExtraParam]

A list of name-value pairs for parameters required by certain top-level domains.

zipCode :: Maybe Text

The zip or postal code of the contact's address.

contactType :: Maybe ContactType

Indicates whether the contact is a person, company, association, or public organization. Note the following:

If you specify a value other than PERSON, you must also specify a value for OrganizationName.

For some TLDs, the privacy protection available depends on the value that you specify for Contact Type. For the privacy protection settings for your TLD, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide

For .es domains, if you specify PERSON, you must specify INDIVIDUAL for the value of ES_LEGAL_FORM.

fax :: Maybe Text

Fax number of the contact.

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code]". For example, a US phone number might appear as "+1.1234567890".

city :: Maybe Text

The city of the contact's address.

state :: Maybe Text

The state or province of the contact's city.

email :: Maybe Text

Email address of the contact.

lastName :: Maybe Text

Last name of contact.

firstName :: Maybe Text

First name of contact.

countryCode :: Maybe CountryCode

Code for the country of the contact's address.

addressLine2 :: Maybe Text

Second line of contact's address, if any.

newContactDetail :: ContactDetail

Create a value of ContactDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneNumber:ContactDetail', contactDetail_phoneNumber - The phone number of the contact.

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code>]". For example, a US phone number might appear as "+1.1234567890".

$sel:organizationName:ContactDetail', contactDetail_organizationName - Name of the organization for contact types other than PERSON.

$sel:addressLine1:ContactDetail', contactDetail_addressLine1 - First line of the contact's address.

$sel:extraParams:ContactDetail', contactDetail_extraParams - A list of name-value pairs for parameters required by certain top-level domains.

$sel:zipCode:ContactDetail', contactDetail_zipCode - The zip or postal code of the contact's address.

$sel:contactType:ContactDetail', contactDetail_contactType - Indicates whether the contact is a person, company, association, or public organization. Note the following:

If you specify a value other than PERSON, you must also specify a value for OrganizationName.

For some TLDs, the privacy protection available depends on the value that you specify for Contact Type. For the privacy protection settings for your TLD, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide

For .es domains, if you specify PERSON, you must specify INDIVIDUAL for the value of ES_LEGAL_FORM.

$sel:fax:ContactDetail', contactDetail_fax - Fax number of the contact.

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code]". For example, a US phone number might appear as "+1.1234567890".

$sel:city:ContactDetail', contactDetail_city - The city of the contact's address.

$sel:state:ContactDetail', contactDetail_state - The state or province of the contact's city.

$sel:email:ContactDetail', contactDetail_email - Email address of the contact.

$sel:lastName:ContactDetail', contactDetail_lastName - Last name of contact.

$sel:firstName:ContactDetail', contactDetail_firstName - First name of contact.

$sel:countryCode:ContactDetail', contactDetail_countryCode - Code for the country of the contact's address.

$sel:addressLine2:ContactDetail', contactDetail_addressLine2 - Second line of contact's address, if any.

contactDetail_phoneNumber :: Lens' ContactDetail (Maybe Text)

The phone number of the contact.

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code>]". For example, a US phone number might appear as "+1.1234567890".

contactDetail_organizationName :: Lens' ContactDetail (Maybe Text)

Name of the organization for contact types other than PERSON.

contactDetail_addressLine1 :: Lens' ContactDetail (Maybe Text)

First line of the contact's address.

contactDetail_extraParams :: Lens' ContactDetail (Maybe [ExtraParam])

A list of name-value pairs for parameters required by certain top-level domains.

contactDetail_zipCode :: Lens' ContactDetail (Maybe Text)

The zip or postal code of the contact's address.

contactDetail_contactType :: Lens' ContactDetail (Maybe ContactType)

Indicates whether the contact is a person, company, association, or public organization. Note the following:

If you specify a value other than PERSON, you must also specify a value for OrganizationName.

For some TLDs, the privacy protection available depends on the value that you specify for Contact Type. For the privacy protection settings for your TLD, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide

For .es domains, if you specify PERSON, you must specify INDIVIDUAL for the value of ES_LEGAL_FORM.

contactDetail_fax :: Lens' ContactDetail (Maybe Text)

Fax number of the contact.

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code]". For example, a US phone number might appear as "+1.1234567890".

contactDetail_city :: Lens' ContactDetail (Maybe Text)

The city of the contact's address.

contactDetail_state :: Lens' ContactDetail (Maybe Text)

The state or province of the contact's city.

contactDetail_email :: Lens' ContactDetail (Maybe Text)

Email address of the contact.

contactDetail_lastName :: Lens' ContactDetail (Maybe Text)

Last name of contact.

contactDetail_firstName :: Lens' ContactDetail (Maybe Text)

First name of contact.

contactDetail_countryCode :: Lens' ContactDetail (Maybe CountryCode)

Code for the country of the contact's address.

contactDetail_addressLine2 :: Lens' ContactDetail (Maybe Text)

Second line of contact's address, if any.

DomainSuggestion

data DomainSuggestion

Information about one suggested domain name.

See: newDomainSuggestion smart constructor.

Constructors

DomainSuggestion'

Fields

availability :: Maybe Text

Whether the domain name is available for registering.

You can register only the domains that are designated as AVAILABLE.

Valid values:

AVAILABLE

The domain name is available.

AVAILABLE_RESERVED

The domain name is reserved under specific conditions.

AVAILABLE_PREORDER

The domain name is available and can be preordered.

DONT_KNOW

The TLD registry didn't reply with a definitive answer about whether the domain name is available. Route 53 can return this response for a variety of reasons, for example, the registry is performing maintenance. Try again later.

PENDING

The TLD registry didn't return a response in the expected amount of time. When the response is delayed, it usually takes just a few extra seconds. You can resubmit the request immediately.

RESERVED

The domain name has been reserved for another person or organization.

UNAVAILABLE

The domain name is not available.

UNAVAILABLE_PREMIUM

The domain name is not available.

UNAVAILABLE_RESTRICTED

The domain name is forbidden.

domainName :: Maybe Text

A suggested domain name.

newDomainSuggestion :: DomainSuggestion

Create a value of DomainSuggestion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availability:DomainSuggestion', domainSuggestion_availability - Whether the domain name is available for registering.

You can register only the domains that are designated as AVAILABLE.

Valid values:

AVAILABLE

The domain name is available.

AVAILABLE_RESERVED

The domain name is reserved under specific conditions.

AVAILABLE_PREORDER

The domain name is available and can be preordered.

DONT_KNOW

The TLD registry didn't reply with a definitive answer about whether the domain name is available. Route 53 can return this response for a variety of reasons, for example, the registry is performing maintenance. Try again later.

PENDING

The TLD registry didn't return a response in the expected amount of time. When the response is delayed, it usually takes just a few extra seconds. You can resubmit the request immediately.

RESERVED

The domain name has been reserved for another person or organization.

UNAVAILABLE

The domain name is not available.

UNAVAILABLE_PREMIUM

The domain name is not available.

UNAVAILABLE_RESTRICTED

The domain name is forbidden.

$sel:domainName:DomainSuggestion', domainSuggestion_domainName - A suggested domain name.

domainSuggestion_availability :: Lens' DomainSuggestion (Maybe Text)

Whether the domain name is available for registering.

You can register only the domains that are designated as AVAILABLE.

Valid values:

AVAILABLE

The domain name is available.

AVAILABLE_RESERVED

The domain name is reserved under specific conditions.

AVAILABLE_PREORDER

The domain name is available and can be preordered.

DONT_KNOW

The TLD registry didn't reply with a definitive answer about whether the domain name is available. Route 53 can return this response for a variety of reasons, for example, the registry is performing maintenance. Try again later.

PENDING

The TLD registry didn't return a response in the expected amount of time. When the response is delayed, it usually takes just a few extra seconds. You can resubmit the request immediately.

RESERVED

The domain name has been reserved for another person or organization.

UNAVAILABLE

The domain name is not available.

UNAVAILABLE_PREMIUM

The domain name is not available.

UNAVAILABLE_RESTRICTED

The domain name is forbidden.

domainSuggestion_domainName :: Lens' DomainSuggestion (Maybe Text)

A suggested domain name.

DomainSummary

data DomainSummary

Summary information about one domain.

See: newDomainSummary smart constructor.

Constructors

DomainSummary'

Fields

expiry :: Maybe POSIX

Expiration date of the domain in Unix time format and Coordinated Universal Time (UTC).

autoRenew :: Maybe Bool

Indicates whether the domain is automatically renewed upon expiration.

transferLock :: Maybe Bool

Indicates whether a domain is locked from unauthorized transfer to another party.

domainName :: Text

The name of the domain that the summary information applies to.

newDomainSummary

Arguments

:: Text
$sel:domainName:DomainSummary'

-> DomainSummary

Create a value of DomainSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expiry:DomainSummary', domainSummary_expiry - Expiration date of the domain in Unix time format and Coordinated Universal Time (UTC).

$sel:autoRenew:DomainSummary', domainSummary_autoRenew - Indicates whether the domain is automatically renewed upon expiration.

$sel:transferLock:DomainSummary', domainSummary_transferLock - Indicates whether a domain is locked from unauthorized transfer to another party.

$sel:domainName:DomainSummary', domainSummary_domainName - The name of the domain that the summary information applies to.

domainSummary_expiry :: Lens' DomainSummary (Maybe UTCTime)

Expiration date of the domain in Unix time format and Coordinated Universal Time (UTC).

domainSummary_autoRenew :: Lens' DomainSummary (Maybe Bool)

Indicates whether the domain is automatically renewed upon expiration.

domainSummary_transferLock :: Lens' DomainSummary (Maybe Bool)

Indicates whether a domain is locked from unauthorized transfer to another party.

domainSummary_domainName :: Lens' DomainSummary Text

The name of the domain that the summary information applies to.

DomainTransferability

data DomainTransferability

A complex type that contains information about whether the specified domain can be transferred to Route 53.

See: newDomainTransferability smart constructor.

Constructors

DomainTransferability'

Fields

transferable :: Maybe Transferable

newDomainTransferability :: DomainTransferability

Create a value of DomainTransferability with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transferable:DomainTransferability', domainTransferability_transferable - Undocumented member.

domainTransferability_transferable :: Lens' DomainTransferability (Maybe Transferable)

Undocumented member.

ExtraParam

data ExtraParam

ExtraParam includes the following elements.

See: newExtraParam smart constructor.

Constructors

ExtraParam'

Fields

name :: ExtraParamName

The name of an additional parameter that is required by a top-level domain. Here are the top-level domains that require additional parameters and the names of the parameters that they require:

.com.au and .net.au

- AU_ID_NUMBER

AU_ID_TYPE

Valid values include the following:

ABN (Australian business number)

ACN (Australian company number)

TM (Trademark number)

.ca

- BRAND_NUMBER

CA_BUSINESS_ENTITY_TYPE

Valid values include the following:

BANK (Bank)

COMMERCIAL_COMPANY (Commercial company)

COMPANY (Company)

COOPERATION (Cooperation)

COOPERATIVE (Cooperative)

COOPRIX (Cooprix)

CORP (Corporation)

CREDIT_UNION (Credit union)

FOMIA (Federation of mutual insurance associations)

INC (Incorporated)

LTD (Limited)

LTEE (Limitée)

LLC (Limited liability corporation)

LLP (Limited liability partnership)

LTE (Lte.)

MBA (Mutual benefit association)

MIC (Mutual insurance company)

NFP (Not-for-profit corporation)

SA (S.A.)

SAVINGS_COMPANY (Savings company)

SAVINGS_UNION (Savings union)

SARL (Société à responsabilité limitée)

TRUST (Trust)

ULC (Unlimited liability corporation)

CA_LEGAL_TYPE

When ContactType is PERSON, valid values include the following:

ABO (Aboriginal Peoples indigenous to Canada)

CCT (Canadian citizen)

LGR (Legal Representative of a Canadian Citizen or Permanent Resident)

RES (Permanent resident of Canada)

When ContactType is a value other than PERSON, valid values include the following:

ASS (Canadian unincorporated association)

CCO (Canadian corporation)

EDU (Canadian educational institution)

GOV (Government or government entity in Canada)

HOP (Canadian Hospital)

INB (Indian Band recognized by the Indian Act of Canada)

LAM (Canadian Library, Archive, or Museum)

MAJ (Her/His Majesty the Queen/King)

OMK (Official mark registered in Canada)

PLT (Canadian Political Party)

PRT (Partnership Registered in Canada)

TDM (Trademark registered in Canada)

TRD (Canadian Trade Union)

TRS (Trust established in Canada)

.es

- ES_IDENTIFICATION
Specify the applicable value:

For contacts inside Spain: Enter your passport ID.

For contacts outside of Spain: Enter the VAT identification number for the company.

For .es domains, the value of ContactType must be PERSON.

ES_IDENTIFICATION_TYPE

Valid values include the following:

DNI_AND_NIF (For Spanish contacts)

NIE (For foreigners with legal residence)

OTHER (For contacts outside of Spain)

ES_LEGAL_FORM

Valid values include the following:

ASSOCIATION

CENTRAL_GOVERNMENT_BODY

CIVIL_SOCIETY

COMMUNITY_OF_OWNERS

COMMUNITY_PROPERTY

CONSULATE

COOPERATIVE

DESIGNATION_OF_ORIGIN_SUPERVISORY_COUNCIL

ECONOMIC_INTEREST_GROUP

EMBASSY

ENTITY_MANAGING_NATURAL_AREAS

FARM_PARTNERSHIP

FOUNDATION

GENERAL_AND_LIMITED_PARTNERSHIP

GENERAL_PARTNERSHIP

INDIVIDUAL

LIMITED_COMPANY

LOCAL_AUTHORITY

LOCAL_PUBLIC_ENTITY

MUTUAL_INSURANCE_COMPANY

NATIONAL_PUBLIC_ENTITY

ORDER_OR_RELIGIOUS_INSTITUTION

OTHERS (Only for contacts outside of Spain)

POLITICAL_PARTY

PROFESSIONAL_ASSOCIATION

PUBLIC_LAW_ASSOCIATION

PUBLIC_LIMITED_COMPANY

REGIONAL_GOVERNMENT_BODY

REGIONAL_PUBLIC_ENTITY

SAVINGS_BANK

SPANISH_OFFICE

SPORTS_ASSOCIATION

SPORTS_FEDERATION

SPORTS_LIMITED_COMPANY

TEMPORARY_ALLIANCE_OF_ENTERPRISES

TRADE_UNION

WORKER_OWNED_COMPANY

WORKER_OWNED_LIMITED_COMPANY

.fi

- BIRTH_DATE_IN_YYYY_MM_DD

FI_BUSINESS_NUMBER

FI_ID_NUMBER

FI_NATIONALITY

Valid values include the following:

FINNISH

NOT_FINNISH

FI_ORGANIZATION_TYPE

Valid values include the following:

COMPANY

CORPORATION

GOVERNMENT

INSTITUTION

POLITICAL_PARTY

PUBLIC_COMMUNITY

TOWNSHIP

.fr

- BIRTH_CITY

BIRTH_COUNTRY

BIRTH_DATE_IN_YYYY_MM_DD

BIRTH_DEPARTMENT: Specify the INSEE code that corresponds with the department where the contact was born. If the contact was born somewhere other than France or its overseas departments, specify 99. For more information, including a list of departments and the corresponding INSEE numbers, see the Wikipedia entry Departments of France.

BRAND_NUMBER

.it

- IT_NATIONALITY

IT_PIN

IT_REGISTRANT_ENTITY_TYPE

Valid values include the following:

FOREIGNERS

FREELANCE_WORKERS (Freelance workers and professionals)

ITALIAN_COMPANIES (Italian companies and one-person companies)

NON_PROFIT_ORGANIZATIONS

OTHER_SUBJECTS

PUBLIC_ORGANIZATIONS

.ru

- BIRTH_DATE_IN_YYYY_MM_DD

RU_PASSPORT_DATA

.se

- BIRTH_COUNTRY

SE_ID_NUMBER

.sg

- SG_ID_NUMBER

.co.uk, .me.uk, and .org.uk

- UK_CONTACT_TYPE
Valid values include the following:

CRC (UK Corporation by Royal Charter)

FCORP (Non-UK Corporation)

FIND (Non-UK Individual, representing self)

FOTHER (Non-UK Entity that does not fit into any other category)

GOV (UK Government Body)

IND (UK Individual (representing self))

IP (UK Industrial/Provident Registered Company)

LLP (UK Limited Liability Partnership)

LTD (UK Limited Company)

OTHER (UK Entity that does not fit into any other category)

PLC (UK Public Limited Company)

PTNR (UK Partnership)

RCHAR (UK Registered Charity)

SCH (UK School)

STAT (UK Statutory Body)

STRA (UK Sole Trader)

UK_COMPANY_NUMBER

In addition, many TLDs require a VAT_NUMBER.

value :: Text

The value that corresponds with the name of an extra parameter.

newExtraParam

Arguments

:: ExtraParamName
$sel:name:ExtraParam'

-> Text
$sel:value:ExtraParam'

-> ExtraParam

Create a value of ExtraParam with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ExtraParam', extraParam_name - The name of an additional parameter that is required by a top-level domain. Here are the top-level domains that require additional parameters and the names of the parameters that they require:

.com.au and .net.au

- AU_ID_NUMBER

AU_ID_TYPE

Valid values include the following:

ABN (Australian business number)

ACN (Australian company number)

TM (Trademark number)

.ca

- BRAND_NUMBER

CA_BUSINESS_ENTITY_TYPE

Valid values include the following:

BANK (Bank)

COMMERCIAL_COMPANY (Commercial company)

COMPANY (Company)

COOPERATION (Cooperation)

COOPERATIVE (Cooperative)

COOPRIX (Cooprix)

CORP (Corporation)

CREDIT_UNION (Credit union)

FOMIA (Federation of mutual insurance associations)

INC (Incorporated)

LTD (Limited)

LTEE (Limitée)

LLC (Limited liability corporation)

LLP (Limited liability partnership)

LTE (Lte.)

MBA (Mutual benefit association)

MIC (Mutual insurance company)

NFP (Not-for-profit corporation)

SA (S.A.)

SAVINGS_COMPANY (Savings company)

SAVINGS_UNION (Savings union)

SARL (Société à responsabilité limitée)

TRUST (Trust)

ULC (Unlimited liability corporation)

CA_LEGAL_TYPE

When ContactType is PERSON, valid values include the following:

ABO (Aboriginal Peoples indigenous to Canada)

CCT (Canadian citizen)

LGR (Legal Representative of a Canadian Citizen or Permanent Resident)

RES (Permanent resident of Canada)

When ContactType is a value other than PERSON, valid values include the following:

ASS (Canadian unincorporated association)

CCO (Canadian corporation)

EDU (Canadian educational institution)

GOV (Government or government entity in Canada)

HOP (Canadian Hospital)

INB (Indian Band recognized by the Indian Act of Canada)

LAM (Canadian Library, Archive, or Museum)

MAJ (Her/His Majesty the Queen/King)

OMK (Official mark registered in Canada)

PLT (Canadian Political Party)

PRT (Partnership Registered in Canada)

TDM (Trademark registered in Canada)

TRD (Canadian Trade Union)

TRS (Trust established in Canada)

.es

- ES_IDENTIFICATION
Specify the applicable value:

For contacts inside Spain: Enter your passport ID.

For contacts outside of Spain: Enter the VAT identification number for the company.

For .es domains, the value of ContactType must be PERSON.

ES_IDENTIFICATION_TYPE

Valid values include the following:

DNI_AND_NIF (For Spanish contacts)

NIE (For foreigners with legal residence)

OTHER (For contacts outside of Spain)

ES_LEGAL_FORM

Valid values include the following:

ASSOCIATION

CENTRAL_GOVERNMENT_BODY

CIVIL_SOCIETY

COMMUNITY_OF_OWNERS

COMMUNITY_PROPERTY

CONSULATE

COOPERATIVE

DESIGNATION_OF_ORIGIN_SUPERVISORY_COUNCIL

ECONOMIC_INTEREST_GROUP

EMBASSY

ENTITY_MANAGING_NATURAL_AREAS

FARM_PARTNERSHIP

FOUNDATION

GENERAL_AND_LIMITED_PARTNERSHIP

GENERAL_PARTNERSHIP

INDIVIDUAL

LIMITED_COMPANY

LOCAL_AUTHORITY

LOCAL_PUBLIC_ENTITY

MUTUAL_INSURANCE_COMPANY

NATIONAL_PUBLIC_ENTITY

ORDER_OR_RELIGIOUS_INSTITUTION

OTHERS (Only for contacts outside of Spain)

POLITICAL_PARTY

PROFESSIONAL_ASSOCIATION

PUBLIC_LAW_ASSOCIATION

PUBLIC_LIMITED_COMPANY

REGIONAL_GOVERNMENT_BODY

REGIONAL_PUBLIC_ENTITY

SAVINGS_BANK

SPANISH_OFFICE

SPORTS_ASSOCIATION

SPORTS_FEDERATION

SPORTS_LIMITED_COMPANY

TEMPORARY_ALLIANCE_OF_ENTERPRISES

TRADE_UNION

WORKER_OWNED_COMPANY

WORKER_OWNED_LIMITED_COMPANY

.fi

- BIRTH_DATE_IN_YYYY_MM_DD

FI_BUSINESS_NUMBER

FI_ID_NUMBER

FI_NATIONALITY

Valid values include the following:

FINNISH

NOT_FINNISH

FI_ORGANIZATION_TYPE

Valid values include the following:

COMPANY

CORPORATION

GOVERNMENT

INSTITUTION

POLITICAL_PARTY

PUBLIC_COMMUNITY

TOWNSHIP

.fr

- BIRTH_CITY

BIRTH_COUNTRY

BIRTH_DATE_IN_YYYY_MM_DD

BIRTH_DEPARTMENT: Specify the INSEE code that corresponds with the department where the contact was born. If the contact was born somewhere other than France or its overseas departments, specify 99. For more information, including a list of departments and the corresponding INSEE numbers, see the Wikipedia entry Departments of France.

BRAND_NUMBER

.it

- IT_NATIONALITY

IT_PIN

IT_REGISTRANT_ENTITY_TYPE

Valid values include the following:

FOREIGNERS

FREELANCE_WORKERS (Freelance workers and professionals)

ITALIAN_COMPANIES (Italian companies and one-person companies)

NON_PROFIT_ORGANIZATIONS

OTHER_SUBJECTS

PUBLIC_ORGANIZATIONS

.ru

- BIRTH_DATE_IN_YYYY_MM_DD

RU_PASSPORT_DATA

.se

- BIRTH_COUNTRY

SE_ID_NUMBER

.sg

- SG_ID_NUMBER

.co.uk, .me.uk, and .org.uk

- UK_CONTACT_TYPE
Valid values include the following:

CRC (UK Corporation by Royal Charter)

FCORP (Non-UK Corporation)

FIND (Non-UK Individual, representing self)

FOTHER (Non-UK Entity that does not fit into any other category)

GOV (UK Government Body)

IND (UK Individual (representing self))

IP (UK Industrial/Provident Registered Company)

LLP (UK Limited Liability Partnership)

LTD (UK Limited Company)

OTHER (UK Entity that does not fit into any other category)

PLC (UK Public Limited Company)

PTNR (UK Partnership)

RCHAR (UK Registered Charity)

SCH (UK School)

STAT (UK Statutory Body)

STRA (UK Sole Trader)

UK_COMPANY_NUMBER

In addition, many TLDs require a VAT_NUMBER.

$sel:value:ExtraParam', extraParam_value - The value that corresponds with the name of an extra parameter.

extraParam_name :: Lens' ExtraParam ExtraParamName

The name of an additional parameter that is required by a top-level domain. Here are the top-level domains that require additional parameters and the names of the parameters that they require:

.com.au and .net.au

- AU_ID_NUMBER

AU_ID_TYPE

Valid values include the following:

ABN (Australian business number)

ACN (Australian company number)

TM (Trademark number)

.ca

- BRAND_NUMBER

CA_BUSINESS_ENTITY_TYPE

Valid values include the following:

BANK (Bank)

COMMERCIAL_COMPANY (Commercial company)

COMPANY (Company)

COOPERATION (Cooperation)

COOPERATIVE (Cooperative)

COOPRIX (Cooprix)

CORP (Corporation)

CREDIT_UNION (Credit union)

FOMIA (Federation of mutual insurance associations)

INC (Incorporated)

LTD (Limited)

LTEE (Limitée)

LLC (Limited liability corporation)

LLP (Limited liability partnership)

LTE (Lte.)

MBA (Mutual benefit association)

MIC (Mutual insurance company)

NFP (Not-for-profit corporation)

SA (S.A.)

SAVINGS_COMPANY (Savings company)

SAVINGS_UNION (Savings union)

SARL (Société à responsabilité limitée)

TRUST (Trust)

ULC (Unlimited liability corporation)

CA_LEGAL_TYPE

When ContactType is PERSON, valid values include the following:

ABO (Aboriginal Peoples indigenous to Canada)

CCT (Canadian citizen)

LGR (Legal Representative of a Canadian Citizen or Permanent Resident)

RES (Permanent resident of Canada)

When ContactType is a value other than PERSON, valid values include the following:

ASS (Canadian unincorporated association)

CCO (Canadian corporation)

EDU (Canadian educational institution)

GOV (Government or government entity in Canada)

HOP (Canadian Hospital)

INB (Indian Band recognized by the Indian Act of Canada)

LAM (Canadian Library, Archive, or Museum)

MAJ (Her/His Majesty the Queen/King)

OMK (Official mark registered in Canada)

PLT (Canadian Political Party)

PRT (Partnership Registered in Canada)

TDM (Trademark registered in Canada)

TRD (Canadian Trade Union)

TRS (Trust established in Canada)

.es

- ES_IDENTIFICATION
Specify the applicable value:

For contacts inside Spain: Enter your passport ID.

For contacts outside of Spain: Enter the VAT identification number for the company.

For .es domains, the value of ContactType must be PERSON.

ES_IDENTIFICATION_TYPE

Valid values include the following:

DNI_AND_NIF (For Spanish contacts)

NIE (For foreigners with legal residence)

OTHER (For contacts outside of Spain)

ES_LEGAL_FORM

Valid values include the following:

ASSOCIATION

CENTRAL_GOVERNMENT_BODY

CIVIL_SOCIETY

COMMUNITY_OF_OWNERS

COMMUNITY_PROPERTY

CONSULATE

COOPERATIVE

DESIGNATION_OF_ORIGIN_SUPERVISORY_COUNCIL

ECONOMIC_INTEREST_GROUP

EMBASSY

ENTITY_MANAGING_NATURAL_AREAS

FARM_PARTNERSHIP

FOUNDATION

GENERAL_AND_LIMITED_PARTNERSHIP

GENERAL_PARTNERSHIP

INDIVIDUAL

LIMITED_COMPANY

LOCAL_AUTHORITY

LOCAL_PUBLIC_ENTITY

MUTUAL_INSURANCE_COMPANY

NATIONAL_PUBLIC_ENTITY

ORDER_OR_RELIGIOUS_INSTITUTION

OTHERS (Only for contacts outside of Spain)

POLITICAL_PARTY

PROFESSIONAL_ASSOCIATION

PUBLIC_LAW_ASSOCIATION

PUBLIC_LIMITED_COMPANY

REGIONAL_GOVERNMENT_BODY

REGIONAL_PUBLIC_ENTITY

SAVINGS_BANK

SPANISH_OFFICE

SPORTS_ASSOCIATION

SPORTS_FEDERATION

SPORTS_LIMITED_COMPANY

TEMPORARY_ALLIANCE_OF_ENTERPRISES

TRADE_UNION

WORKER_OWNED_COMPANY

WORKER_OWNED_LIMITED_COMPANY

.fi

- BIRTH_DATE_IN_YYYY_MM_DD

FI_BUSINESS_NUMBER

FI_ID_NUMBER

FI_NATIONALITY

Valid values include the following:

FINNISH

NOT_FINNISH

FI_ORGANIZATION_TYPE

Valid values include the following:

COMPANY

CORPORATION

GOVERNMENT

INSTITUTION

POLITICAL_PARTY

PUBLIC_COMMUNITY

TOWNSHIP

.fr

- BIRTH_CITY

BIRTH_COUNTRY

BIRTH_DATE_IN_YYYY_MM_DD

BIRTH_DEPARTMENT: Specify the INSEE code that corresponds with the department where the contact was born. If the contact was born somewhere other than France or its overseas departments, specify 99. For more information, including a list of departments and the corresponding INSEE numbers, see the Wikipedia entry Departments of France.

BRAND_NUMBER

.it

- IT_NATIONALITY

IT_PIN

IT_REGISTRANT_ENTITY_TYPE

Valid values include the following:

FOREIGNERS

FREELANCE_WORKERS (Freelance workers and professionals)

ITALIAN_COMPANIES (Italian companies and one-person companies)

NON_PROFIT_ORGANIZATIONS

OTHER_SUBJECTS

PUBLIC_ORGANIZATIONS

.ru

- BIRTH_DATE_IN_YYYY_MM_DD

RU_PASSPORT_DATA

.se

- BIRTH_COUNTRY

SE_ID_NUMBER

.sg

- SG_ID_NUMBER

.co.uk, .me.uk, and .org.uk

- UK_CONTACT_TYPE
Valid values include the following:

CRC (UK Corporation by Royal Charter)

FCORP (Non-UK Corporation)

FIND (Non-UK Individual, representing self)

FOTHER (Non-UK Entity that does not fit into any other category)

GOV (UK Government Body)

IND (UK Individual (representing self))

IP (UK Industrial/Provident Registered Company)

LLP (UK Limited Liability Partnership)

LTD (UK Limited Company)

OTHER (UK Entity that does not fit into any other category)

PLC (UK Public Limited Company)

PTNR (UK Partnership)

RCHAR (UK Registered Charity)

SCH (UK School)

STAT (UK Statutory Body)

STRA (UK Sole Trader)

UK_COMPANY_NUMBER

In addition, many TLDs require a VAT_NUMBER.

extraParam_value :: Lens' ExtraParam Text

The value that corresponds with the name of an extra parameter.

Nameserver

data Nameserver

Nameserver includes the following elements.

See: newNameserver smart constructor.

Constructors

Nameserver'

Fields

glueIps :: Maybe [Text]

Glue IP address of a name server entry. Glue IP addresses are required only when the name of the name server is a subdomain of the domain. For example, if your domain is example.com and the name server for the domain is ns.example.com, you need to specify the IP address for ns.example.com.

Constraints: The list can contain only one IPv4 and one IPv6 address.

name :: Text

The fully qualified host name of the name server.

Constraint: Maximum 255 characters

newNameserver

Arguments

:: Text
$sel:name:Nameserver'

-> Nameserver

Create a value of Nameserver with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:glueIps:Nameserver', nameserver_glueIps - Glue IP address of a name server entry. Glue IP addresses are required only when the name of the name server is a subdomain of the domain. For example, if your domain is example.com and the name server for the domain is ns.example.com, you need to specify the IP address for ns.example.com.

Constraints: The list can contain only one IPv4 and one IPv6 address.

$sel:name:Nameserver', nameserver_name - The fully qualified host name of the name server.

Constraint: Maximum 255 characters

nameserver_glueIps :: Lens' Nameserver (Maybe [Text])

Glue IP address of a name server entry. Glue IP addresses are required only when the name of the name server is a subdomain of the domain. For example, if your domain is example.com and the name server for the domain is ns.example.com, you need to specify the IP address for ns.example.com.

Constraints: The list can contain only one IPv4 and one IPv6 address.

nameserver_name :: Lens' Nameserver Text

The fully qualified host name of the name server.

Constraint: Maximum 255 characters

OperationSummary

data OperationSummary

OperationSummary includes the following elements.

See: newOperationSummary smart constructor.

Constructors

OperationSummary'

Fields

operationId :: Text

Identifier returned to track the requested action.

status :: OperationStatus

The current status of the requested operation in the system.

type' :: OperationType

Type of the action requested.

submittedDate :: POSIX

The date when the request was submitted.

newOperationSummary

Arguments

:: Text
$sel:operationId:OperationSummary'

-> OperationStatus
$sel:status:OperationSummary'

-> OperationType
$sel:type':OperationSummary'

-> UTCTime
$sel:submittedDate:OperationSummary'

-> OperationSummary

Create a value of OperationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationId:OperationSummary', operationSummary_operationId - Identifier returned to track the requested action.

$sel:status:OperationSummary', operationSummary_status - The current status of the requested operation in the system.

$sel:type':OperationSummary', operationSummary_type - Type of the action requested.

$sel:submittedDate:OperationSummary', operationSummary_submittedDate - The date when the request was submitted.

operationSummary_operationId :: Lens' OperationSummary Text

Identifier returned to track the requested action.

operationSummary_status :: Lens' OperationSummary OperationStatus

The current status of the requested operation in the system.

operationSummary_type :: Lens' OperationSummary OperationType

Type of the action requested.

operationSummary_submittedDate :: Lens' OperationSummary UTCTime

The date when the request was submitted.

Tag

data Tag

Each tag includes the following elements.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Maybe Text

The key (name) of a tag.

Valid values: A-Z, a-z, 0-9, space, ".:/=+\-@"

Constraints: Each key can be 1-128 characters long.

value :: Maybe Text

The value of a tag.

Valid values: A-Z, a-z, 0-9, space, ".:/=+\-@"

Constraints: Each value can be 0-256 characters long.

newTag :: Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key (name) of a tag.

Valid values: A-Z, a-z, 0-9, space, ".:/=+\-@"

Constraints: Each key can be 1-128 characters long.

$sel:value:Tag', tag_value - The value of a tag.

Valid values: A-Z, a-z, 0-9, space, ".:/=+\-@"

Constraints: Each value can be 0-256 characters long.

tag_key :: Lens' Tag (Maybe Text)

The key (name) of a tag.

Valid values: A-Z, a-z, 0-9, space, ".:/=+\-@"

Constraints: Each key can be 1-128 characters long.

tag_value :: Lens' Tag (Maybe Text)

The value of a tag.

Valid values: A-Z, a-z, 0-9, space, ".:/=+\-@"

Constraints: Each value can be 0-256 characters long.