Network.AWS.Route53Domains.Types.BillingRecord

Description

Documentation

data BillingRecord

Information for one billing record.

See: newBillingRecord smart constructor.

Constructors

BillingRecord'

Fields

invoiceId :: Maybe Text

The ID of the invoice that is associated with the billing record.

operation :: Maybe OperationType

The operation that you were charged for.

billDate :: Maybe POSIX

The date that the operation was billed, in Unix format.

domainName :: Maybe Text

The name of the domain that the billing record applies to. If the domain name contains characters other than a-z, 0-9, and - (hyphen), such as an internationalized domain name, then this value is in Punycode. For more information, see DNS Domain Name Format in the Amazon Route 53 Developer Guide.

price :: Maybe Double

The price that you were charged for the operation, in US dollars.

Example value: 12.0

newBillingRecord :: BillingRecord

Create a value of BillingRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:invoiceId:BillingRecord', billingRecord_invoiceId - The ID of the invoice that is associated with the billing record.

$sel:operation:BillingRecord', billingRecord_operation - The operation that you were charged for.

$sel:billDate:BillingRecord', billingRecord_billDate - The date that the operation was billed, in Unix format.

$sel:domainName:BillingRecord', billingRecord_domainName - The name of the domain that the billing record applies to. If the domain name contains characters other than a-z, 0-9, and - (hyphen), such as an internationalized domain name, then this value is in Punycode. For more information, see DNS Domain Name Format in the Amazon Route 53 Developer Guide.

$sel:price:BillingRecord', billingRecord_price - The price that you were charged for the operation, in US dollars.

Example value: 12.0

billingRecord_invoiceId :: Lens' BillingRecord (Maybe Text)

The ID of the invoice that is associated with the billing record.

billingRecord_operation :: Lens' BillingRecord (Maybe OperationType)

The operation that you were charged for.

billingRecord_billDate :: Lens' BillingRecord (Maybe UTCTime)

The date that the operation was billed, in Unix format.

billingRecord_domainName :: Lens' BillingRecord (Maybe Text)

The name of the domain that the billing record applies to. If the domain name contains characters other than a-z, 0-9, and - (hyphen), such as an internationalized domain name, then this value is in Punycode. For more information, see DNS Domain Name Format in the Amazon Route 53 Developer Guide.

billingRecord_price :: Lens' BillingRecord (Maybe Double)

The price that you were charged for the operation, in US dollars.

Example value: 12.0