Network.AWS.Route53Domains.UpdateDomainContact

Description

This operation updates the contact information for a particular domain. You must specify information for at least one contact: registrant, administrator, or technical.

If the update is successful, this method returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

Creating a Request

data UpdateDomainContact

The UpdateDomainContact request includes the following elements.

See: newUpdateDomainContact smart constructor.

Constructors

UpdateDomainContact'

Fields

registrantContact :: Maybe (Sensitive ContactDetail)

Provides detailed contact information.

techContact :: Maybe (Sensitive ContactDetail)

Provides detailed contact information.

adminContact :: Maybe (Sensitive ContactDetail)

Provides detailed contact information.

domainName :: Text

The name of the domain that you want to update contact information for.

newUpdateDomainContact

Arguments

:: Text
$sel:domainName:UpdateDomainContact'

-> UpdateDomainContact

Create a value of UpdateDomainContact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registrantContact:UpdateDomainContact', updateDomainContact_registrantContact - Provides detailed contact information.

$sel:techContact:UpdateDomainContact', updateDomainContact_techContact - Provides detailed contact information.

$sel:adminContact:UpdateDomainContact', updateDomainContact_adminContact - Provides detailed contact information.

$sel:domainName:UpdateDomainContact', updateDomainContact_domainName - The name of the domain that you want to update contact information for.

Request Lenses

updateDomainContact_registrantContact :: Lens' UpdateDomainContact (Maybe ContactDetail)

Provides detailed contact information.

updateDomainContact_techContact :: Lens' UpdateDomainContact (Maybe ContactDetail)

Provides detailed contact information.

updateDomainContact_adminContact :: Lens' UpdateDomainContact (Maybe ContactDetail)

Provides detailed contact information.

updateDomainContact_domainName :: Lens' UpdateDomainContact Text

The name of the domain that you want to update contact information for.

Destructuring the Response

data UpdateDomainContactResponse

The UpdateDomainContact response includes the following element.

See: newUpdateDomainContactResponse smart constructor.

Constructors

UpdateDomainContactResponse'

Fields

httpStatus :: Int

The response's http status code.

operationId :: Text

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

newUpdateDomainContactResponse

Arguments

:: Int
$sel:httpStatus:UpdateDomainContactResponse'

-> Text
$sel:operationId:UpdateDomainContactResponse'

-> UpdateDomainContactResponse

Create a value of UpdateDomainContactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateDomainContactResponse', updateDomainContactResponse_httpStatus - The response's http status code.

$sel:operationId:UpdateDomainContactResponse', updateDomainContactResponse_operationId - Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

Response Lenses

updateDomainContactResponse_httpStatus :: Lens' UpdateDomainContactResponse Int

The response's http status code.

updateDomainContactResponse_operationId :: Lens' UpdateDomainContactResponse Text

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.