Network.AWS.Route53Domains.UpdateDomainContactPrivacy

Description

This operation updates the specified domain contact's privacy setting. When privacy protection is enabled, contact information such as email address is replaced either with contact information for Amazon Registrar (for .com, .net, and .org domains) or with contact information for our registrar associate, Gandi.

This operation affects only the contact information for the specified contact type (registrant, administrator, or tech). If the request succeeds, Amazon Route 53 returns an operation ID that you can use with GetOperationDetail to track the progress and completion of the action. If the request doesn't complete successfully, the domain registrant will be notified by email.

By disabling the privacy service via API, you consent to the publication of the contact information provided for this domain via the public WHOIS database. You certify that you are the registrant of this domain name and have the authority to make this decision. You may withdraw your consent at any time by enabling privacy protection using either UpdateDomainContactPrivacy or the Route 53 console. Enabling privacy protection removes the contact information provided for this domain from the WHOIS database. For more information on our privacy practices, see https://aws.amazon.com/privacy/.

Creating a Request

data UpdateDomainContactPrivacy

The UpdateDomainContactPrivacy request includes the following elements.

See: newUpdateDomainContactPrivacy smart constructor.

Constructors

UpdateDomainContactPrivacy'

Fields

adminPrivacy :: Maybe Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the admin contact.

techPrivacy :: Maybe Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the technical contact.

registrantPrivacy :: Maybe Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the registrant contact (domain owner).

domainName :: Text

The name of the domain that you want to update the privacy setting for.

newUpdateDomainContactPrivacy

Arguments

:: Text
$sel:domainName:UpdateDomainContactPrivacy'

-> UpdateDomainContactPrivacy

Create a value of UpdateDomainContactPrivacy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:adminPrivacy:UpdateDomainContactPrivacy', updateDomainContactPrivacy_adminPrivacy - Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the admin contact.

$sel:techPrivacy:UpdateDomainContactPrivacy', updateDomainContactPrivacy_techPrivacy - Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the technical contact.

$sel:registrantPrivacy:UpdateDomainContactPrivacy', updateDomainContactPrivacy_registrantPrivacy - Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the registrant contact (domain owner).

$sel:domainName:UpdateDomainContactPrivacy', updateDomainContactPrivacy_domainName - The name of the domain that you want to update the privacy setting for.

Request Lenses

updateDomainContactPrivacy_adminPrivacy :: Lens' UpdateDomainContactPrivacy (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the admin contact.

updateDomainContactPrivacy_techPrivacy :: Lens' UpdateDomainContactPrivacy (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the technical contact.

updateDomainContactPrivacy_registrantPrivacy :: Lens' UpdateDomainContactPrivacy (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the registrant contact (domain owner).

updateDomainContactPrivacy_domainName :: Lens' UpdateDomainContactPrivacy Text

The name of the domain that you want to update the privacy setting for.

Destructuring the Response

data UpdateDomainContactPrivacyResponse

The UpdateDomainContactPrivacy response includes the following element.

See: newUpdateDomainContactPrivacyResponse smart constructor.

Constructors

UpdateDomainContactPrivacyResponse'

Fields

httpStatus :: Int

The response's http status code.

operationId :: Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail.

newUpdateDomainContactPrivacyResponse

Arguments

:: Int
$sel:httpStatus:UpdateDomainContactPrivacyResponse'

-> Text
$sel:operationId:UpdateDomainContactPrivacyResponse'

-> UpdateDomainContactPrivacyResponse

Create a value of UpdateDomainContactPrivacyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateDomainContactPrivacyResponse', updateDomainContactPrivacyResponse_httpStatus - The response's http status code.

$sel:operationId:UpdateDomainContactPrivacyResponse', updateDomainContactPrivacyResponse_operationId - Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail.

Response Lenses

updateDomainContactPrivacyResponse_httpStatus :: Lens' UpdateDomainContactPrivacyResponse Int

The response's http status code.

updateDomainContactPrivacyResponse_operationId :: Lens' UpdateDomainContactPrivacyResponse Text

Identifier for tracking the progress of the request. To use this ID to query the operation status, use GetOperationDetail.