Network.AWS.Route53RecoveryCluster.Types

Description

Service Configuration

defaultService :: Service

API version 2019-12-02 of the Amazon Route53 Recovery Cluster SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

There was a validation error on the request.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You don't have sufficient permissions to query the routing control state.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

There was a conflict with this request. Try again.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied because of request throttling.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

There was an unexpected error during processing of the request.

_EndpointTemporarilyUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The cluster endpoint isn't available. Try another cluster endpoint.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The request references a routing control that was not found.

RoutingControlState

newtype RoutingControlState

Constructors

RoutingControlState'

Fields

fromRoutingControlState :: Text

Bundled Patterns

pattern RoutingControlState_Off :: RoutingControlState

pattern RoutingControlState_On :: RoutingControlState

UpdateRoutingControlStateEntry

data UpdateRoutingControlStateEntry

A routing control state.

See: newUpdateRoutingControlStateEntry smart constructor.

Constructors

UpdateRoutingControlStateEntry'

Fields

routingControlArn :: Text

The Amazon Resource Number (ARN) for the routing control state entry.

routingControlState :: RoutingControlState

The routing control state in a set of routing control state entries.

newUpdateRoutingControlStateEntry

Arguments

:: Text
$sel:routingControlArn:UpdateRoutingControlStateEntry'

-> RoutingControlState
$sel:routingControlState:UpdateRoutingControlStateEntry'

-> UpdateRoutingControlStateEntry

Create a value of UpdateRoutingControlStateEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:routingControlArn:UpdateRoutingControlStateEntry', updateRoutingControlStateEntry_routingControlArn - The Amazon Resource Number (ARN) for the routing control state entry.

$sel:routingControlState:UpdateRoutingControlStateEntry', updateRoutingControlStateEntry_routingControlState - The routing control state in a set of routing control state entries.

updateRoutingControlStateEntry_routingControlArn :: Lens' UpdateRoutingControlStateEntry Text

The Amazon Resource Number (ARN) for the routing control state entry.

updateRoutingControlStateEntry_routingControlState :: Lens' UpdateRoutingControlStateEntry RoutingControlState

The routing control state in a set of routing control state entries.