Network.AWS.Route53RecoveryControlConfig

Description

Derived from API version 2020-11-02 of the AWS service descriptions, licensed under Apache 2.0.

Recovery Control Configuration API Reference for Amazon Route 53 Application Recovery Controller

Service Configuration

defaultService :: Service

API version 2020-11-02 of the Amazon Route53 Recovery Control Config SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Route53RecoveryControlConfig.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

400 response - Multiple causes. For example, you might have a malformed query string and input parameter might be out of range, or you used parameters together incorrectly.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

403 response - AccessDeniedException. You do not have sufficient access to perform this action.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

409 response - ConflictException.

ServiceQuotaExceededException

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

402 response

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

429 response - ThrottlingException.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

500 response - InternalServiceError. Temporary service error. Retry the request.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

404 response - The query string contains a syntax error or resource not found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

ClusterCreated

newClusterCreated :: Wait DescribeCluster

Polls DescribeCluster every 5 seconds until a successful state is reached. An error is returned after 26 failed checks.

RoutingControlCreated

newRoutingControlCreated :: Wait DescribeRoutingControl

Polls DescribeRoutingControl every 5 seconds until a successful state is reached. An error is returned after 26 failed checks.

RoutingControlDeleted

newRoutingControlDeleted :: Wait DescribeRoutingControl

Polls DescribeRoutingControl every 5 seconds until a successful state is reached. An error is returned after 26 failed checks.

ControlPanelCreated

newControlPanelCreated :: Wait DescribeControlPanel

Polls DescribeControlPanel every 5 seconds until a successful state is reached. An error is returned after 26 failed checks.

ControlPanelDeleted

newControlPanelDeleted :: Wait DescribeControlPanel

Polls DescribeControlPanel every 5 seconds until a successful state is reached. An error is returned after 26 failed checks.

ClusterDeleted

newClusterDeleted :: Wait DescribeCluster

Polls DescribeCluster every 5 seconds until a successful state is reached. An error is returned after 26 failed checks.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DescribeControlPanel

data DescribeControlPanel

See: newDescribeControlPanel smart constructor.

Constructors

DescribeControlPanel' Text

newDescribeControlPanel

Arguments

:: Text
$sel:controlPanelArn:DescribeControlPanel'

-> DescribeControlPanel

Create a value of DescribeControlPanel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:controlPanelArn:DescribeControlPanel', describeControlPanel_controlPanelArn - The Amazon Resource Name (ARN) of the control panel that you're getting details for.

data DescribeControlPanelResponse

See: newDescribeControlPanelResponse smart constructor.

Constructors

DescribeControlPanelResponse' (Maybe ControlPanel) Int

newDescribeControlPanelResponse

Arguments

:: Int
$sel:httpStatus:DescribeControlPanelResponse'

-> DescribeControlPanelResponse

Create a value of DescribeControlPanelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:controlPanel:DescribeControlPanelResponse', describeControlPanelResponse_controlPanel - Information about the control panel.

$sel:httpStatus:DescribeControlPanelResponse', describeControlPanelResponse_httpStatus - The response's http status code.

CreateRoutingControl

data CreateRoutingControl

The details of the routing control that you're creating.

See: newCreateRoutingControl smart constructor.

Constructors

CreateRoutingControl' (Maybe Text) (Maybe Text) Text Text

newCreateRoutingControl

Arguments

:: Text
$sel:clusterArn:CreateRoutingControl'

-> Text
$sel:routingControlName:CreateRoutingControl'

-> CreateRoutingControl

Create a value of CreateRoutingControl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:controlPanelArn:CreateRoutingControl', createRoutingControl_controlPanelArn - The Amazon Resource Name (ARN) of the control panel that includes the routing control.

$sel:clientToken:CreateRoutingControl', createRoutingControl_clientToken - Unique client idempotency token.

$sel:clusterArn:CreateRoutingControl', createRoutingControl_clusterArn - The Amazon Resource Name (ARN) of the cluster that includes the routing control.

$sel:routingControlName:CreateRoutingControl', createRoutingControl_routingControlName - The name of the routing control.

data CreateRoutingControlResponse

See: newCreateRoutingControlResponse smart constructor.

Constructors

CreateRoutingControlResponse' (Maybe RoutingControl) Int

newCreateRoutingControlResponse

Arguments

:: Int
$sel:httpStatus:CreateRoutingControlResponse'

-> CreateRoutingControlResponse

Create a value of CreateRoutingControlResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:routingControl:CreateRoutingControlResponse', createRoutingControlResponse_routingControl - The routing control that is created.

$sel:httpStatus:CreateRoutingControlResponse', createRoutingControlResponse_httpStatus - The response's http status code.

DescribeCluster

data DescribeCluster

See: newDescribeCluster smart constructor.

Constructors

DescribeCluster' Text

newDescribeCluster

Arguments

:: Text
$sel:clusterArn:DescribeCluster'

-> DescribeCluster

Create a value of DescribeCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:DescribeCluster', describeCluster_clusterArn - The Amazon Resource Name (ARN) of the cluster that you're getting details for.

data DescribeClusterResponse

See: newDescribeClusterResponse smart constructor.

Constructors

DescribeClusterResponse' (Maybe Cluster) Int

newDescribeClusterResponse

Arguments

:: Int
$sel:httpStatus:DescribeClusterResponse'

-> DescribeClusterResponse

Create a value of DescribeClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:DescribeClusterResponse', describeClusterResponse_cluster - The cluster for the DescribeCluster request.

$sel:httpStatus:DescribeClusterResponse', describeClusterResponse_httpStatus - The response's http status code.

DeleteRoutingControl

data DeleteRoutingControl

See: newDeleteRoutingControl smart constructor.

Constructors

DeleteRoutingControl' Text

newDeleteRoutingControl

Arguments

:: Text
$sel:routingControlArn:DeleteRoutingControl'

-> DeleteRoutingControl

Create a value of DeleteRoutingControl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:routingControlArn:DeleteRoutingControl', deleteRoutingControl_routingControlArn - The Amazon Resource Name (ARN) of the routing control that you're deleting.

data DeleteRoutingControlResponse

See: newDeleteRoutingControlResponse smart constructor.

Constructors

DeleteRoutingControlResponse' Int

newDeleteRoutingControlResponse

Arguments

:: Int
$sel:httpStatus:DeleteRoutingControlResponse'

-> DeleteRoutingControlResponse

Create a value of DeleteRoutingControlResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteRoutingControlResponse', deleteRoutingControlResponse_httpStatus - The response's http status code.

UpdateRoutingControl

data UpdateRoutingControl

The details of the routing control that you're updating.

See: newUpdateRoutingControl smart constructor.

Constructors

UpdateRoutingControl' Text Text

newUpdateRoutingControl

Arguments

:: Text
$sel:routingControlName:UpdateRoutingControl'

-> Text
$sel:routingControlArn:UpdateRoutingControl'

-> UpdateRoutingControl

Create a value of UpdateRoutingControl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:routingControlName:UpdateRoutingControl', updateRoutingControl_routingControlName - The name of the routing control.

$sel:routingControlArn:UpdateRoutingControl', updateRoutingControl_routingControlArn - The Amazon Resource Name (ARN) of the routing control.

data UpdateRoutingControlResponse

See: newUpdateRoutingControlResponse smart constructor.

Constructors

UpdateRoutingControlResponse' (Maybe RoutingControl) Int

newUpdateRoutingControlResponse

Arguments

:: Int
$sel:httpStatus:UpdateRoutingControlResponse'

-> UpdateRoutingControlResponse

Create a value of UpdateRoutingControlResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:routingControl:UpdateRoutingControlResponse', updateRoutingControlResponse_routingControl - The routing control that was updated.

$sel:httpStatus:UpdateRoutingControlResponse', updateRoutingControlResponse_httpStatus - The response's http status code.

CreateControlPanel

data CreateControlPanel

The details of the control panel that you're creating.

See: newCreateControlPanel smart constructor.

Constructors

CreateControlPanel' (Maybe Text) Text Text

newCreateControlPanel

Arguments

:: Text
$sel:clusterArn:CreateControlPanel'

-> Text
$sel:controlPanelName:CreateControlPanel'

-> CreateControlPanel

Create a value of CreateControlPanel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateControlPanel', createControlPanel_clientToken - Unique client idempotency token.

$sel:clusterArn:CreateControlPanel', createControlPanel_clusterArn - The Amazon Resource Name (ARN) of the cluster for the control panel.

$sel:controlPanelName:CreateControlPanel', createControlPanel_controlPanelName - The name of the control panel.

data CreateControlPanelResponse

See: newCreateControlPanelResponse smart constructor.

Constructors

CreateControlPanelResponse' (Maybe ControlPanel) Int

newCreateControlPanelResponse

Arguments

:: Int
$sel:httpStatus:CreateControlPanelResponse'

-> CreateControlPanelResponse

Create a value of CreateControlPanelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:controlPanel:CreateControlPanelResponse', createControlPanelResponse_controlPanel - Information about a control panel.

$sel:httpStatus:CreateControlPanelResponse', createControlPanelResponse_httpStatus - The response's http status code.

UpdateControlPanel

data UpdateControlPanel

The details of the control panel that you're updating.

See: newUpdateControlPanel smart constructor.

Constructors

UpdateControlPanel' Text Text

newUpdateControlPanel

Arguments

:: Text
$sel:controlPanelArn:UpdateControlPanel'

-> Text
$sel:controlPanelName:UpdateControlPanel'

-> UpdateControlPanel

Create a value of UpdateControlPanel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:controlPanelArn:UpdateControlPanel', updateControlPanel_controlPanelArn - The Amazon Resource Name (ARN) of the control panel.

$sel:controlPanelName:UpdateControlPanel', updateControlPanel_controlPanelName - The name of the control panel.

data UpdateControlPanelResponse

See: newUpdateControlPanelResponse smart constructor.

Constructors

UpdateControlPanelResponse' (Maybe ControlPanel) Int

newUpdateControlPanelResponse

Arguments

:: Int
$sel:httpStatus:UpdateControlPanelResponse'

-> UpdateControlPanelResponse

Create a value of UpdateControlPanelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:controlPanel:UpdateControlPanelResponse', updateControlPanelResponse_controlPanel - The control panel to update.

$sel:httpStatus:UpdateControlPanelResponse', updateControlPanelResponse_httpStatus - The response's http status code.

DeleteControlPanel

data DeleteControlPanel

See: newDeleteControlPanel smart constructor.

Constructors

DeleteControlPanel' Text

newDeleteControlPanel

Arguments

:: Text
$sel:controlPanelArn:DeleteControlPanel'

-> DeleteControlPanel

Create a value of DeleteControlPanel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:controlPanelArn:DeleteControlPanel', deleteControlPanel_controlPanelArn - The Amazon Resource Name (ARN) of the control panel that you're deleting.

data DeleteControlPanelResponse

See: newDeleteControlPanelResponse smart constructor.

Constructors

DeleteControlPanelResponse' Int

newDeleteControlPanelResponse

Arguments

:: Int
$sel:httpStatus:DeleteControlPanelResponse'

-> DeleteControlPanelResponse

Create a value of DeleteControlPanelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteControlPanelResponse', deleteControlPanelResponse_httpStatus - The response's http status code.

DeleteCluster

data DeleteCluster

See: newDeleteCluster smart constructor.

Constructors

DeleteCluster' Text

newDeleteCluster

Arguments

:: Text
$sel:clusterArn:DeleteCluster'

-> DeleteCluster

Create a value of DeleteCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:DeleteCluster', deleteCluster_clusterArn - The Amazon Resource Name (ARN) of the cluster that you're deleting.

data DeleteClusterResponse

See: newDeleteClusterResponse smart constructor.

Constructors

DeleteClusterResponse' Int

newDeleteClusterResponse

Arguments

:: Int
$sel:httpStatus:DeleteClusterResponse'

-> DeleteClusterResponse

Create a value of DeleteClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteClusterResponse', deleteClusterResponse_httpStatus - The response's http status code.

CreateSafetyRule

data CreateSafetyRule

The request body that you include when you create a safety rule.

See: newCreateSafetyRule smart constructor.

Constructors

CreateSafetyRule' (Maybe NewAssertionRule) (Maybe Text) (Maybe NewGatingRule)

newCreateSafetyRule :: CreateSafetyRule

Create a value of CreateSafetyRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assertionRule:CreateSafetyRule', createSafetyRule_assertionRule - Undocumented member.

$sel:clientToken:CreateSafetyRule', createSafetyRule_clientToken - Unique client idempotency token.

$sel:gatingRule:CreateSafetyRule', createSafetyRule_gatingRule - Undocumented member.

data CreateSafetyRuleResponse

See: newCreateSafetyRuleResponse smart constructor.

Constructors

CreateSafetyRuleResponse' (Maybe AssertionRule) (Maybe GatingRule) Int

newCreateSafetyRuleResponse

Arguments

:: Int
$sel:httpStatus:CreateSafetyRuleResponse'

-> CreateSafetyRuleResponse

Create a value of CreateSafetyRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assertionRule:CreateSafetyRule', createSafetyRuleResponse_assertionRule - Undocumented member.

$sel:gatingRule:CreateSafetyRule', createSafetyRuleResponse_gatingRule - Undocumented member.

$sel:httpStatus:CreateSafetyRuleResponse', createSafetyRuleResponse_httpStatus - The response's http status code.

CreateCluster

data CreateCluster

Creates a cluster.

See: newCreateCluster smart constructor.

Constructors

CreateCluster' (Maybe Text) Text

newCreateCluster

Arguments

:: Text
$sel:clusterName:CreateCluster'

-> CreateCluster

Create a value of CreateCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateCluster', createCluster_clientToken - Unique client idempotency token.

$sel:clusterName:CreateCluster', createCluster_clusterName - The name of the cluster.

data CreateClusterResponse

See: newCreateClusterResponse smart constructor.

Constructors

CreateClusterResponse' (Maybe Cluster) Int

newCreateClusterResponse

Arguments

:: Int
$sel:httpStatus:CreateClusterResponse'

-> CreateClusterResponse

Create a value of CreateClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:CreateClusterResponse', createClusterResponse_cluster - The cluster that was created.

$sel:httpStatus:CreateClusterResponse', createClusterResponse_httpStatus - The response's http status code.

ListAssociatedRoute53HealthChecks

data ListAssociatedRoute53HealthChecks

See: newListAssociatedRoute53HealthChecks smart constructor.

Constructors

ListAssociatedRoute53HealthChecks' (Maybe Text) (Maybe Natural) Text

newListAssociatedRoute53HealthChecks

Arguments

:: Text
$sel:routingControlArn:ListAssociatedRoute53HealthChecks'

-> ListAssociatedRoute53HealthChecks

Create a value of ListAssociatedRoute53HealthChecks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssociatedRoute53HealthChecks', listAssociatedRoute53HealthChecks_nextToken - The token that identifies which batch of results you want to see.

$sel:maxResults:ListAssociatedRoute53HealthChecks', listAssociatedRoute53HealthChecks_maxResults - The number of objects that you want to return with this call.

$sel:routingControlArn:ListAssociatedRoute53HealthChecks', listAssociatedRoute53HealthChecks_routingControlArn - The Amazon Resource Name (ARN) of the routing control that you're getting details for.

data ListAssociatedRoute53HealthChecksResponse

See: newListAssociatedRoute53HealthChecksResponse smart constructor.

Constructors

ListAssociatedRoute53HealthChecksResponse' (Maybe Text) (Maybe [Text]) Int

newListAssociatedRoute53HealthChecksResponse

Arguments

:: Int
$sel:httpStatus:ListAssociatedRoute53HealthChecksResponse'

-> ListAssociatedRoute53HealthChecksResponse

Create a value of ListAssociatedRoute53HealthChecksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssociatedRoute53HealthChecks', listAssociatedRoute53HealthChecksResponse_nextToken - The token that identifies which batch of results you want to see.

$sel:healthCheckIds:ListAssociatedRoute53HealthChecksResponse', listAssociatedRoute53HealthChecksResponse_healthCheckIds - Identifiers for the health checks.

$sel:httpStatus:ListAssociatedRoute53HealthChecksResponse', listAssociatedRoute53HealthChecksResponse_httpStatus - The response's http status code.

DescribeSafetyRule

data DescribeSafetyRule

See: newDescribeSafetyRule smart constructor.

Constructors

DescribeSafetyRule' Text

newDescribeSafetyRule

Arguments

:: Text
$sel:safetyRuleArn:DescribeSafetyRule'

-> DescribeSafetyRule

Create a value of DescribeSafetyRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:safetyRuleArn:DescribeSafetyRule', describeSafetyRule_safetyRuleArn - The request body that you include when you update a safety rule.

data DescribeSafetyRuleResponse

See: newDescribeSafetyRuleResponse smart constructor.

Constructors

DescribeSafetyRuleResponse' (Maybe AssertionRule) (Maybe GatingRule) Int

newDescribeSafetyRuleResponse

Arguments

:: Int
$sel:httpStatus:DescribeSafetyRuleResponse'

-> DescribeSafetyRuleResponse

Create a value of DescribeSafetyRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assertionRule:DescribeSafetyRuleResponse', describeSafetyRuleResponse_assertionRule - Undocumented member.

$sel:gatingRule:DescribeSafetyRuleResponse', describeSafetyRuleResponse_gatingRule - Undocumented member.

$sel:httpStatus:DescribeSafetyRuleResponse', describeSafetyRuleResponse_httpStatus - The response's http status code.

ListRoutingControls

data ListRoutingControls

See: newListRoutingControls smart constructor.

Constructors

ListRoutingControls' (Maybe Text) (Maybe Natural) Text

newListRoutingControls

Arguments

:: Text
$sel:controlPanelArn:ListRoutingControls'

-> ListRoutingControls

Create a value of ListRoutingControls with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRoutingControls', listRoutingControls_nextToken - The token that identifies which batch of results you want to see.

$sel:maxResults:ListRoutingControls', listRoutingControls_maxResults - The number of objects that you want to return with this call.

$sel:controlPanelArn:ListRoutingControls', listRoutingControls_controlPanelArn - The Amazon Resource Name (ARN) of the control panel that you're getting routing control details for.

data ListRoutingControlsResponse

See: newListRoutingControlsResponse smart constructor.

Constructors

ListRoutingControlsResponse' (Maybe Text) (Maybe [RoutingControl]) Int

newListRoutingControlsResponse

Arguments

:: Int
$sel:httpStatus:ListRoutingControlsResponse'

-> ListRoutingControlsResponse

Create a value of ListRoutingControlsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRoutingControls', listRoutingControlsResponse_nextToken - The token that identifies which batch of results you want to see.

$sel:routingControls:ListRoutingControlsResponse', listRoutingControlsResponse_routingControls - An array of routing controls.

$sel:httpStatus:ListRoutingControlsResponse', listRoutingControlsResponse_httpStatus - The response's http status code.

ListControlPanels

data ListControlPanels

See: newListControlPanels smart constructor.

Constructors

ListControlPanels' (Maybe Text) (Maybe Text) (Maybe Natural)

newListControlPanels :: ListControlPanels

Create a value of ListControlPanels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:ListControlPanels', listControlPanels_clusterArn - The Amazon Resource Name (ARN) of a cluster.

$sel:nextToken:ListControlPanels', listControlPanels_nextToken - The token that identifies which batch of results you want to see.

$sel:maxResults:ListControlPanels', listControlPanels_maxResults - The number of objects that you want to return with this call.

data ListControlPanelsResponse

See: newListControlPanelsResponse smart constructor.

Constructors

ListControlPanelsResponse' (Maybe Text) (Maybe [ControlPanel]) Int

newListControlPanelsResponse

Arguments

:: Int
$sel:httpStatus:ListControlPanelsResponse'

-> ListControlPanelsResponse

Create a value of ListControlPanelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListControlPanels', listControlPanelsResponse_nextToken - The token that identifies which batch of results you want to see.

$sel:controlPanels:ListControlPanelsResponse', listControlPanelsResponse_controlPanels - The result of a successful ListControlPanel request.

$sel:httpStatus:ListControlPanelsResponse', listControlPanelsResponse_httpStatus - The response's http status code.

UpdateSafetyRule

data UpdateSafetyRule

See: newUpdateSafetyRule smart constructor.

Constructors

UpdateSafetyRule' (Maybe GatingRuleUpdate) (Maybe AssertionRuleUpdate)

newUpdateSafetyRule :: UpdateSafetyRule

Create a value of UpdateSafetyRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatingRuleUpdate:UpdateSafetyRule', updateSafetyRule_gatingRuleUpdate - Undocumented member.

$sel:assertionRuleUpdate:UpdateSafetyRule', updateSafetyRule_assertionRuleUpdate - Undocumented member.

data UpdateSafetyRuleResponse

See: newUpdateSafetyRuleResponse smart constructor.

Constructors

UpdateSafetyRuleResponse' (Maybe AssertionRule) (Maybe GatingRule) Int

newUpdateSafetyRuleResponse

Arguments

:: Int
$sel:httpStatus:UpdateSafetyRuleResponse'

-> UpdateSafetyRuleResponse

Create a value of UpdateSafetyRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assertionRule:UpdateSafetyRuleResponse', updateSafetyRuleResponse_assertionRule - Undocumented member.

$sel:gatingRule:UpdateSafetyRuleResponse', updateSafetyRuleResponse_gatingRule - Undocumented member.

$sel:httpStatus:UpdateSafetyRuleResponse', updateSafetyRuleResponse_httpStatus - The response's http status code.

DeleteSafetyRule

data DeleteSafetyRule

See: newDeleteSafetyRule smart constructor.

Constructors

DeleteSafetyRule' Text

newDeleteSafetyRule

Arguments

:: Text
$sel:safetyRuleArn:DeleteSafetyRule'

-> DeleteSafetyRule

Create a value of DeleteSafetyRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:safetyRuleArn:DeleteSafetyRule', deleteSafetyRule_safetyRuleArn - The request body that you include when you update a safety rule.

data DeleteSafetyRuleResponse

See: newDeleteSafetyRuleResponse smart constructor.

Constructors

DeleteSafetyRuleResponse' Int

newDeleteSafetyRuleResponse

Arguments

:: Int
$sel:httpStatus:DeleteSafetyRuleResponse'

-> DeleteSafetyRuleResponse

Create a value of DeleteSafetyRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteSafetyRuleResponse', deleteSafetyRuleResponse_httpStatus - The response's http status code.

ListClusters

data ListClusters

See: newListClusters smart constructor.

Constructors

ListClusters' (Maybe Text) (Maybe Natural)

newListClusters :: ListClusters

Create a value of ListClusters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListClusters', listClusters_nextToken - The token that identifies which batch of results you want to see.

$sel:maxResults:ListClusters', listClusters_maxResults - The number of objects that you want to return with this call.

data ListClustersResponse

See: newListClustersResponse smart constructor.

Constructors

ListClustersResponse' (Maybe Text) (Maybe [Cluster]) Int

newListClustersResponse

Arguments

:: Int
$sel:httpStatus:ListClustersResponse'

-> ListClustersResponse

Create a value of ListClustersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListClusters', listClustersResponse_nextToken - The token that identifies which batch of results you want to see.

$sel:clusters:ListClustersResponse', listClustersResponse_clusters - An array of the clusters in an account.

$sel:httpStatus:ListClustersResponse', listClustersResponse_httpStatus - The response's http status code.

ListSafetyRules

data ListSafetyRules

See: newListSafetyRules smart constructor.

Constructors

ListSafetyRules' (Maybe Text) (Maybe Natural) Text

newListSafetyRules

Arguments

:: Text
$sel:controlPanelArn:ListSafetyRules'

-> ListSafetyRules

Create a value of ListSafetyRules with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSafetyRules', listSafetyRules_nextToken - The token that identifies which batch of results you want to see.

$sel:maxResults:ListSafetyRules', listSafetyRules_maxResults - The number of objects that you want to return with this call.

$sel:controlPanelArn:ListSafetyRules', listSafetyRules_controlPanelArn - The Amazon Resource Name (ARN) of the control panel that you're getting details for.

data ListSafetyRulesResponse

See: newListSafetyRulesResponse smart constructor.

Constructors

ListSafetyRulesResponse' (Maybe Text) (Maybe [Rule]) Int

newListSafetyRulesResponse

Arguments

:: Int
$sel:httpStatus:ListSafetyRulesResponse'

-> ListSafetyRulesResponse

Create a value of ListSafetyRulesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSafetyRules', listSafetyRulesResponse_nextToken - The token that identifies which batch of results you want to see.

$sel:safetyRules:ListSafetyRulesResponse', listSafetyRulesResponse_safetyRules - The list of safety rules in a control panel.

$sel:httpStatus:ListSafetyRulesResponse', listSafetyRulesResponse_httpStatus - The response's http status code.

DescribeRoutingControl

data DescribeRoutingControl

See: newDescribeRoutingControl smart constructor.

Constructors

DescribeRoutingControl' Text

newDescribeRoutingControl

Arguments

:: Text
$sel:routingControlArn:DescribeRoutingControl'

-> DescribeRoutingControl

Create a value of DescribeRoutingControl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:routingControlArn:DescribeRoutingControl', describeRoutingControl_routingControlArn - The Amazon Resource Name (ARN) of the routing control that you're getting details for.

data DescribeRoutingControlResponse

See: newDescribeRoutingControlResponse smart constructor.

Constructors

DescribeRoutingControlResponse' (Maybe RoutingControl) Int

newDescribeRoutingControlResponse

Arguments

:: Int
$sel:httpStatus:DescribeRoutingControlResponse'

-> DescribeRoutingControlResponse

Create a value of DescribeRoutingControlResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:routingControl:DescribeRoutingControlResponse', describeRoutingControlResponse_routingControl - Information about the routing control.

$sel:httpStatus:DescribeRoutingControlResponse', describeRoutingControlResponse_httpStatus - The response's http status code.

Types

RuleType

newtype RuleType

An enumerated type that determines how the evaluated rules are processed. RuleType can be one of the following:

ATLEAST - At least N routing controls must be set. You specify N as the Threshold in the rule configuration.

AND - All routing controls must be set. This is a shortcut for "At least N," where N is the total number of controls in the rule.

OR - Any control must be set. This is a shortcut for "At least N," where N is 1.

Constructors

RuleType'

Fields

fromRuleType :: Text

Bundled Patterns

pattern RuleType_AND :: RuleType

pattern RuleType_ATLEAST :: RuleType

pattern RuleType_OR :: RuleType

Status

newtype Status

The deployment status of a resource. Status can be one of the following:

PENDING: Amazon Route 53 Application Recovery Controller is creating the resource.

DEPLOYED: The resource is deployed and ready to use.

PENDING_DELETION: Amazon Route 53 Application Recovery Controller is deleting the resource.

Constructors

Status'

Fields

fromStatus :: Text

Bundled Patterns

pattern Status_DEPLOYED :: Status

pattern Status_PENDING :: Status

pattern Status_PENDING_DELETION :: Status

AssertionRule

data AssertionRule

An assertion rule enforces that, when a routing control state is changed, the criteria set by the rule configuration is met. Otherwise, the change to the routing control is not accepted.

See: newAssertionRule smart constructor.

Constructors

AssertionRule' Status Text Text [Text] RuleConfig Int Text

newAssertionRule

Arguments

:: Status
$sel:status:AssertionRule'

-> Text
$sel:controlPanelArn:AssertionRule'

-> Text
$sel:safetyRuleArn:AssertionRule'

-> RuleConfig
$sel:ruleConfig:AssertionRule'

-> Int
$sel:waitPeriodMs:AssertionRule'

-> Text
$sel:name:AssertionRule'

-> AssertionRule

Create a value of AssertionRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AssertionRule', assertionRule_status - The deployment status of an assertion rule. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

$sel:controlPanelArn:AssertionRule', assertionRule_controlPanelArn - The Amazon Resource Name (ARN) of the control panel.

$sel:safetyRuleArn:AssertionRule', assertionRule_safetyRuleArn - The Amazon Resource Name (ARN) of the assertion rule.

$sel:assertedControls:AssertionRule', assertionRule_assertedControls - The routing controls that are part of transactions that are evaluated to determine if a request to change a routing control state is allowed. For example, you might include three routing controls, one for each of three Amazon Web Services Regions.

$sel:ruleConfig:AssertionRule', assertionRule_ruleConfig - The criteria that you set for specific assertion controls (routing controls) that designate how many controls must be enabled as the result of a transaction. For example, if you have three assertion controls, you might specify atleast 2 for your rule configuration. This means that at least two assertion controls must be enabled, so that at least two Amazon Web Services Regions are enabled.

$sel:waitPeriodMs:AssertionRule', assertionRule_waitPeriodMs - An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

$sel:name:AssertionRule', assertionRule_name - Name of the assertion rule. You can use any non-white space character in the name.

AssertionRuleUpdate

data AssertionRuleUpdate

An update to an assertion rule. You can update the name or the evaluation period (wait period). If you don't specify one of the items to update, the item is unchanged.

See: newAssertionRuleUpdate smart constructor.

Constructors

AssertionRuleUpdate' Text Int Text

newAssertionRuleUpdate

Arguments

:: Text
$sel:safetyRuleArn:AssertionRuleUpdate'

-> Int
$sel:waitPeriodMs:AssertionRuleUpdate'

-> Text
$sel:name:AssertionRuleUpdate'

-> AssertionRuleUpdate

Create a value of AssertionRuleUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:safetyRuleArn:AssertionRuleUpdate', assertionRuleUpdate_safetyRuleArn - The Amazon Resource Name (ARN) of the assertion rule.

$sel:waitPeriodMs:AssertionRuleUpdate', assertionRuleUpdate_waitPeriodMs - An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

$sel:name:AssertionRuleUpdate', assertionRuleUpdate_name - The name of the assertion rule. You can use any non-white space character in the name.

Cluster

data Cluster

A cluster is a set of five consensus-forming Regional endpoints that represent the infrastructure that hosts your routing controls. Typically, you host together on one cluster all of the routing controls for your applications.

See: newCluster smart constructor.

Constructors

Cluster' (Maybe Status) (Maybe Text) (Maybe Text) (Maybe [ClusterEndpoint])

newCluster :: Cluster

Create a value of Cluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Cluster', cluster_status - Deployment status of a resource. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

$sel:clusterArn:Cluster', cluster_clusterArn - The Amazon Resource Name (ARN) of the cluster.

$sel:name:Cluster', cluster_name - The name of the cluster.

$sel:clusterEndpoints:Cluster', cluster_clusterEndpoints - Endpoints for a cluster. Specify one of these endpoints when you want to set or retrieve a routing control state in the cluster.

To get or update the routing control state, see the Amazon Route 53 Application Recovery Controller Cluster (Data Plane) Actions.

ClusterEndpoint

data ClusterEndpoint

A cluster endpoint. Specify an endpoint when you want to set or retrieve a routing control state in the cluster.

See: newClusterEndpoint smart constructor.

Constructors

ClusterEndpoint' (Maybe Text) (Maybe Text)

newClusterEndpoint :: ClusterEndpoint

Create a value of ClusterEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:region:ClusterEndpoint', clusterEndpoint_region - The Amazon Web Services Region for a cluster endpoint.

$sel:endpoint:ClusterEndpoint', clusterEndpoint_endpoint - A cluster endpoint. Specify an endpoint and Amazon Web Services Region when you want to set or retrieve a routing control state in the cluster.

To get or update the routing control state, see the Amazon Route 53 Application Recovery Controller Cluster (Data Plane) Actions.

ControlPanel

data ControlPanel

A control panel represents a group of routing controls that can be changed together in a single transaction.

See: newControlPanel smart constructor.

Constructors

ControlPanel' (Maybe Status) (Maybe Text) (Maybe Text) (Maybe Int) (Maybe Text) (Maybe Bool)

newControlPanel :: ControlPanel

Create a value of ControlPanel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ControlPanel', controlPanel_status - The deployment status of control panel. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

$sel:controlPanelArn:ControlPanel', controlPanel_controlPanelArn - The Amazon Resource Name (ARN) of the control panel.

$sel:clusterArn:ControlPanel', controlPanel_clusterArn - The Amazon Resource Name (ARN) of the cluster that includes the control panel.

$sel:routingControlCount:ControlPanel', controlPanel_routingControlCount - The number of routing controls in the control panel.

$sel:name:ControlPanel', controlPanel_name - The name of the control panel. You can use any non-white space character in the name.

$sel:defaultControlPanel:ControlPanel', controlPanel_defaultControlPanel - A flag that Amazon Route 53 Application Recovery Controller sets to true to designate the default control panel for a cluster. When you create a cluster, Amazon Route 53 Application Recovery Controller creates a control panel, and sets this flag for that control panel. If you create a control panel yourself, this flag is set to false.

GatingRule

data GatingRule

A gating rule verifies that a set of gating controls evaluates as true, based on a rule configuration that you specify. If the gating rule evaluates to true, Amazon Route 53 Application Recovery Controller allows a set of routing control state changes to run and complete against the set of target controls.

See: newGatingRule smart constructor.

Constructors

GatingRule' Status [Text] Text Text [Text] RuleConfig Int Text

newGatingRule

Arguments

:: Status
$sel:status:GatingRule'

-> Text
$sel:controlPanelArn:GatingRule'

-> Text
$sel:safetyRuleArn:GatingRule'

-> RuleConfig
$sel:ruleConfig:GatingRule'

-> Int
$sel:waitPeriodMs:GatingRule'

-> Text
$sel:name:GatingRule'

-> GatingRule

Create a value of GatingRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GatingRule', gatingRule_status - The deployment status of a gating rule. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

$sel:targetControls:GatingRule', gatingRule_targetControls - Routing controls that can only be set or unset if the specified RuleConfig evaluates to true for the specified GatingControls. For example, say you have three gating controls, one for each of three Amazon Web Services Regions. Now you specify ATLEAST 2 as your RuleConfig. With these settings, you can only change (set or unset) the routing controls that you have specified as TargetControls if that rule evaluates to true.

In other words, your ability to change the routing controls that you have specified as TargetControls is gated by the rule that you set for the routing controls in GatingControls.

$sel:controlPanelArn:GatingRule', gatingRule_controlPanelArn - The Amazon Resource Name (ARN) of the control panel.

$sel:safetyRuleArn:GatingRule', gatingRule_safetyRuleArn - The Amazon Resource Name (ARN) of the gating rule.

$sel:gatingControls:GatingRule', gatingRule_gatingControls - The gating controls for the gating rule. That is, routing controls that are evaluated by the rule configuration that you specify.

$sel:ruleConfig:GatingRule', gatingRule_ruleConfig - The criteria that you set for specific gating controls (routing controls) that designates how many controls must be enabled to allow you to change (set or unset) the target controls.

$sel:waitPeriodMs:GatingRule', gatingRule_waitPeriodMs - An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

$sel:name:GatingRule', gatingRule_name - The name for the gating rule.

GatingRuleUpdate

data GatingRuleUpdate

Update to a gating rule. You can update the name or the evaluation period (wait period). If you don't specify one of the items to update, the item is unchanged.

See: newGatingRuleUpdate smart constructor.

Constructors

GatingRuleUpdate' Text Int Text

newGatingRuleUpdate

Arguments

:: Text
$sel:safetyRuleArn:GatingRuleUpdate'

-> Int
$sel:waitPeriodMs:GatingRuleUpdate'

-> Text
$sel:name:GatingRuleUpdate'

-> GatingRuleUpdate

Create a value of GatingRuleUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:safetyRuleArn:GatingRuleUpdate', gatingRuleUpdate_safetyRuleArn - The Amazon Resource Name (ARN) of the gating rule.

$sel:waitPeriodMs:GatingRuleUpdate', gatingRuleUpdate_waitPeriodMs - An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

$sel:name:GatingRuleUpdate', gatingRuleUpdate_name - The name for the gating rule.

NewAssertionRule

data NewAssertionRule

A new assertion rule for a control panel.

See: newNewAssertionRule smart constructor.

Constructors

NewAssertionRule' Text [Text] RuleConfig Int Text

newNewAssertionRule

Arguments

:: Text
$sel:controlPanelArn:NewAssertionRule'

-> RuleConfig
$sel:ruleConfig:NewAssertionRule'

-> Int
$sel:waitPeriodMs:NewAssertionRule'

-> Text
$sel:name:NewAssertionRule'

-> NewAssertionRule

Create a value of NewAssertionRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:controlPanelArn:NewAssertionRule', newAssertionRule_controlPanelArn - The Amazon Resource Name (ARN) for the control panel.

$sel:assertedControls:NewAssertionRule', newAssertionRule_assertedControls - The routing controls that are part of transactions that are evaluated to determine if a request to change a routing control state is allowed. For example, you might include three routing controls, one for each of three Amazon Web Services Regions.

$sel:ruleConfig:NewAssertionRule', newAssertionRule_ruleConfig - The criteria that you set for specific assertion controls (routing controls) that designate how many controls must be enabled as the result of a transaction. For example, if you have three assertion controls, you might specify atleast 2 for your rule configuration. This means that at least two assertion controls must be enabled, so that at least two Amazon Web Services Regions are enabled.

$sel:waitPeriodMs:NewAssertionRule', newAssertionRule_waitPeriodMs - An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

$sel:name:NewAssertionRule', newAssertionRule_name - The name of the assertion rule. You can use any non-white space character in the name.

NewGatingRule

data NewGatingRule

A new gating rule for a control panel.

See: newNewGatingRule smart constructor.

Constructors

NewGatingRule' [Text] Text [Text] RuleConfig Int Text

newNewGatingRule

Arguments

:: Text
$sel:controlPanelArn:NewGatingRule'

-> RuleConfig
$sel:ruleConfig:NewGatingRule'

-> Int
$sel:waitPeriodMs:NewGatingRule'

-> Text
$sel:name:NewGatingRule'

-> NewGatingRule

Create a value of NewGatingRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetControls:NewGatingRule', newGatingRule_targetControls - Routing controls that can only be set or unset if the specified RuleConfig evaluates to true for the specified GatingControls. For example, say you have three gating controls, one for each of three Amazon Web Services Regions. Now you specify AtLeast 2 as your RuleConfig. With these settings, you can only change (set or unset) the routing controls that you have specified as TargetControls if that rule evaluates to true.

In other words, your ability to change the routing controls that you have specified as TargetControls is gated by the rule that you set for the routing controls in GatingControls.

$sel:controlPanelArn:NewGatingRule', newGatingRule_controlPanelArn - The Amazon Resource Name (ARN) of the control panel.

$sel:gatingControls:NewGatingRule', newGatingRule_gatingControls - The gating controls for the new gating rule. That is, routing controls that are evaluated by the rule configuration that you specify.

$sel:ruleConfig:NewGatingRule', newGatingRule_ruleConfig - The criteria that you set for specific gating controls (routing controls) that designates how many controls must be enabled to allow you to change (set or unset) the target controls.

$sel:waitPeriodMs:NewGatingRule', newGatingRule_waitPeriodMs - An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

$sel:name:NewGatingRule', newGatingRule_name - The name for the new gating rule.

RoutingControl

data RoutingControl

A routing control has one of two states: ON and OFF. You can map the routing control state to the state of an Amazon Route 53 health check, which can be used to control traffic routing.

See: newRoutingControl smart constructor.

Constructors

RoutingControl' (Maybe Status) (Maybe Text) (Maybe Text) (Maybe Text)

newRoutingControl :: RoutingControl

Create a value of RoutingControl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:RoutingControl', routingControl_status - The deployment status of a routing control. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

$sel:controlPanelArn:RoutingControl', routingControl_controlPanelArn - The Amazon Resource Name (ARN) of the control panel that includes the routing control.

$sel:name:RoutingControl', routingControl_name - The name of the routing control.

$sel:routingControlArn:RoutingControl', routingControl_routingControlArn - The Amazon Resource Name (ARN) of the routing control.

Rule

data Rule

A safety rule. A safety rule can be an assertion rule or a gating rule.

See: newRule smart constructor.

Constructors

Rule' (Maybe GatingRule) (Maybe AssertionRule)

newRule :: Rule

Create a value of Rule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gating:Rule', rule_gating - A gating rule verifies that a set of gating controls evaluates as true, based on a rule configuration that you specify. If the gating rule evaluates to true, Amazon Route 53 Application Recovery Controller allows a set of routing control state changes to run and complete against the set of target controls.

$sel:assertion:Rule', rule_assertion - An assertion rule enforces that, when a routing control state is changed, the criteria set by the rule configuration is met. Otherwise, the change to the routing control is not accepted.

RuleConfig

data RuleConfig

The rule configuration for an assertion rule. That is, the criteria that you set for specific assertion controls (routing controls) that specify how many controls must be enabled after a transaction completes.

See: newRuleConfig smart constructor.

Constructors

RuleConfig' RuleType Bool Int

newRuleConfig

Arguments

:: RuleType
$sel:type':RuleConfig'

-> Bool
$sel:inverted:RuleConfig'

-> Int
$sel:threshold:RuleConfig'

-> RuleConfig

Create a value of RuleConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':RuleConfig', ruleConfig_type - A rule can be one of the following: ATLEAST, AND, or OR.

$sel:inverted:RuleConfig', ruleConfig_inverted - Logical negation of the rule. If the rule would usually evaluate true, it's evaluated as false, and vice versa.

$sel:threshold:RuleConfig', ruleConfig_threshold - The value of N, when you specify an ATLEAST rule type. That is, Threshold is the number of controls that must be set when you specify an ATLEAST type.