Network.AWS.Route53RecoveryControlConfig.CreateCluster

Description

Create a new cluster. A cluster is a set of redundant Regional endpoints against which you can run API calls to update or get the state of one or more routing controls. Each cluster has a name, status, Amazon Resource Name (ARN), and an array of the five cluster endpoints (one for each supported Amazon Web Services Region) that you can use with API calls to the Amazon Route 53 Application Recovery Controller cluster data plane.

Creating a Request

data CreateCluster

Creates a cluster.

See: newCreateCluster smart constructor.

Constructors

CreateCluster'

Fields

clientToken :: Maybe Text

Unique client idempotency token.

clusterName :: Text

The name of the cluster.

newCreateCluster

Arguments

:: Text
$sel:clusterName:CreateCluster'

-> CreateCluster

Create a value of CreateCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateCluster', createCluster_clientToken - Unique client idempotency token.

$sel:clusterName:CreateCluster', createCluster_clusterName - The name of the cluster.

Request Lenses

createCluster_clientToken :: Lens' CreateCluster (Maybe Text)

Unique client idempotency token.

createCluster_clusterName :: Lens' CreateCluster Text

The name of the cluster.

Destructuring the Response

data CreateClusterResponse

See: newCreateClusterResponse smart constructor.

Constructors

CreateClusterResponse'

Fields

cluster :: Maybe Cluster

The cluster that was created.

httpStatus :: Int

The response's http status code.

newCreateClusterResponse

Arguments

:: Int
$sel:httpStatus:CreateClusterResponse'

-> CreateClusterResponse

Create a value of CreateClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:CreateClusterResponse', createClusterResponse_cluster - The cluster that was created.

$sel:httpStatus:CreateClusterResponse', createClusterResponse_httpStatus - The response's http status code.

Response Lenses

createClusterResponse_cluster :: Lens' CreateClusterResponse (Maybe Cluster)

The cluster that was created.

createClusterResponse_httpStatus :: Lens' CreateClusterResponse Int

The response's http status code.