Network.AWS.Route53RecoveryControlConfig.DescribeControlPanel

Description

Displays details about a control panel.

Creating a Request

data DescribeControlPanel

See: newDescribeControlPanel smart constructor.

Constructors

DescribeControlPanel'

Fields

controlPanelArn :: Text

The Amazon Resource Name (ARN) of the control panel that you're getting details for.

newDescribeControlPanel

Arguments

:: Text
$sel:controlPanelArn:DescribeControlPanel'

-> DescribeControlPanel

Create a value of DescribeControlPanel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:controlPanelArn:DescribeControlPanel', describeControlPanel_controlPanelArn - The Amazon Resource Name (ARN) of the control panel that you're getting details for.

Request Lenses

describeControlPanel_controlPanelArn :: Lens' DescribeControlPanel Text

The Amazon Resource Name (ARN) of the control panel that you're getting details for.

Destructuring the Response

data DescribeControlPanelResponse

See: newDescribeControlPanelResponse smart constructor.

Constructors

DescribeControlPanelResponse'

Fields

controlPanel :: Maybe ControlPanel

Information about the control panel.

httpStatus :: Int

The response's http status code.

newDescribeControlPanelResponse

Arguments

:: Int
$sel:httpStatus:DescribeControlPanelResponse'

-> DescribeControlPanelResponse

Create a value of DescribeControlPanelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:controlPanel:DescribeControlPanelResponse', describeControlPanelResponse_controlPanel - Information about the control panel.

$sel:httpStatus:DescribeControlPanelResponse', describeControlPanelResponse_httpStatus - The response's http status code.

Response Lenses

describeControlPanelResponse_controlPanel :: Lens' DescribeControlPanelResponse (Maybe ControlPanel)

Information about the control panel.

describeControlPanelResponse_httpStatus :: Lens' DescribeControlPanelResponse Int

The response's http status code.