Network.AWS.Route53RecoveryControlConfig.ListRoutingControls

Description

Returns an array of routing controls for a control panel. A routing control is an Amazon Route 53 Application Recovery Controller construct that has one of two states: ON and OFF. You can map the routing control state to the state of an Amazon Route 53 health check, which can be used to control routing.

Creating a Request

data ListRoutingControls

See: newListRoutingControls smart constructor.

Constructors

ListRoutingControls'

Fields

nextToken :: Maybe Text

The token that identifies which batch of results you want to see.

maxResults :: Maybe Natural

The number of objects that you want to return with this call.

controlPanelArn :: Text

The Amazon Resource Name (ARN) of the control panel that you're getting routing control details for.

newListRoutingControls

Arguments

:: Text
$sel:controlPanelArn:ListRoutingControls'

-> ListRoutingControls

Create a value of ListRoutingControls with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRoutingControls', listRoutingControls_nextToken - The token that identifies which batch of results you want to see.

$sel:maxResults:ListRoutingControls', listRoutingControls_maxResults - The number of objects that you want to return with this call.

$sel:controlPanelArn:ListRoutingControls', listRoutingControls_controlPanelArn - The Amazon Resource Name (ARN) of the control panel that you're getting routing control details for.

Request Lenses

listRoutingControls_nextToken :: Lens' ListRoutingControls (Maybe Text)

The token that identifies which batch of results you want to see.

listRoutingControls_maxResults :: Lens' ListRoutingControls (Maybe Natural)

The number of objects that you want to return with this call.

listRoutingControls_controlPanelArn :: Lens' ListRoutingControls Text

The Amazon Resource Name (ARN) of the control panel that you're getting routing control details for.

Destructuring the Response

data ListRoutingControlsResponse

See: newListRoutingControlsResponse smart constructor.

Constructors

ListRoutingControlsResponse'

Fields

nextToken :: Maybe Text

The token that identifies which batch of results you want to see.

routingControls :: Maybe [RoutingControl]

An array of routing controls.

httpStatus :: Int

The response's http status code.

newListRoutingControlsResponse

Arguments

:: Int
$sel:httpStatus:ListRoutingControlsResponse'

-> ListRoutingControlsResponse

Create a value of ListRoutingControlsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRoutingControls', listRoutingControlsResponse_nextToken - The token that identifies which batch of results you want to see.

$sel:routingControls:ListRoutingControlsResponse', listRoutingControlsResponse_routingControls - An array of routing controls.

$sel:httpStatus:ListRoutingControlsResponse', listRoutingControlsResponse_httpStatus - The response's http status code.

Response Lenses

listRoutingControlsResponse_nextToken :: Lens' ListRoutingControlsResponse (Maybe Text)

The token that identifies which batch of results you want to see.

listRoutingControlsResponse_routingControls :: Lens' ListRoutingControlsResponse (Maybe [RoutingControl])

An array of routing controls.

listRoutingControlsResponse_httpStatus :: Lens' ListRoutingControlsResponse Int

The response's http status code.