Network.AWS.Route53RecoveryControlConfig.Types

Description

Service Configuration

defaultService :: Service

API version 2020-11-02 of the Amazon Route53 Recovery Control Config SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

400 response - Multiple causes. For example, you might have a malformed query string and input parameter might be out of range, or you used parameters together incorrectly.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

403 response - AccessDeniedException. You do not have sufficient access to perform this action.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

409 response - ConflictException.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

402 response

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

429 response - ThrottlingException.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

500 response - InternalServiceError. Temporary service error. Retry the request.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

404 response - The query string contains a syntax error or resource not found.

RuleType

newtype RuleType

An enumerated type that determines how the evaluated rules are processed. RuleType can be one of the following:

ATLEAST - At least N routing controls must be set. You specify N as the Threshold in the rule configuration.

AND - All routing controls must be set. This is a shortcut for "At least N," where N is the total number of controls in the rule.

OR - Any control must be set. This is a shortcut for "At least N," where N is 1.

Constructors

RuleType'

Fields

fromRuleType :: Text

Bundled Patterns

pattern RuleType_AND :: RuleType

pattern RuleType_ATLEAST :: RuleType

pattern RuleType_OR :: RuleType

Status

newtype Status

The deployment status of a resource. Status can be one of the following:

PENDING: Amazon Route 53 Application Recovery Controller is creating the resource.

DEPLOYED: The resource is deployed and ready to use.

PENDING_DELETION: Amazon Route 53 Application Recovery Controller is deleting the resource.

Constructors

Status'

Fields

fromStatus :: Text

Bundled Patterns

pattern Status_DEPLOYED :: Status

pattern Status_PENDING :: Status

pattern Status_PENDING_DELETION :: Status

AssertionRule

data AssertionRule

An assertion rule enforces that, when a routing control state is changed, the criteria set by the rule configuration is met. Otherwise, the change to the routing control is not accepted.

See: newAssertionRule smart constructor.

Constructors

AssertionRule'

Fields

status :: Status

The deployment status of an assertion rule. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

controlPanelArn :: Text

The Amazon Resource Name (ARN) of the control panel.

safetyRuleArn :: Text

The Amazon Resource Name (ARN) of the assertion rule.

assertedControls :: [Text]

The routing controls that are part of transactions that are evaluated to determine if a request to change a routing control state is allowed. For example, you might include three routing controls, one for each of three Amazon Web Services Regions.

ruleConfig :: RuleConfig

The criteria that you set for specific assertion controls (routing controls) that designate how many controls must be enabled as the result of a transaction. For example, if you have three assertion controls, you might specify atleast 2 for your rule configuration. This means that at least two assertion controls must be enabled, so that at least two Amazon Web Services Regions are enabled.

waitPeriodMs :: Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

name :: Text

Name of the assertion rule. You can use any non-white space character in the name.

newAssertionRule

Arguments

:: Status
$sel:status:AssertionRule'

-> Text
$sel:controlPanelArn:AssertionRule'

-> Text
$sel:safetyRuleArn:AssertionRule'

-> RuleConfig
$sel:ruleConfig:AssertionRule'

-> Int
$sel:waitPeriodMs:AssertionRule'

-> Text
$sel:name:AssertionRule'

-> AssertionRule

Create a value of AssertionRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AssertionRule', assertionRule_status - The deployment status of an assertion rule. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

$sel:controlPanelArn:AssertionRule', assertionRule_controlPanelArn - The Amazon Resource Name (ARN) of the control panel.

$sel:safetyRuleArn:AssertionRule', assertionRule_safetyRuleArn - The Amazon Resource Name (ARN) of the assertion rule.

$sel:assertedControls:AssertionRule', assertionRule_assertedControls - The routing controls that are part of transactions that are evaluated to determine if a request to change a routing control state is allowed. For example, you might include three routing controls, one for each of three Amazon Web Services Regions.

$sel:ruleConfig:AssertionRule', assertionRule_ruleConfig - The criteria that you set for specific assertion controls (routing controls) that designate how many controls must be enabled as the result of a transaction. For example, if you have three assertion controls, you might specify atleast 2 for your rule configuration. This means that at least two assertion controls must be enabled, so that at least two Amazon Web Services Regions are enabled.

$sel:waitPeriodMs:AssertionRule', assertionRule_waitPeriodMs - An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

$sel:name:AssertionRule', assertionRule_name - Name of the assertion rule. You can use any non-white space character in the name.

assertionRule_status :: Lens' AssertionRule Status

The deployment status of an assertion rule. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

assertionRule_controlPanelArn :: Lens' AssertionRule Text

The Amazon Resource Name (ARN) of the control panel.

assertionRule_safetyRuleArn :: Lens' AssertionRule Text

The Amazon Resource Name (ARN) of the assertion rule.

assertionRule_assertedControls :: Lens' AssertionRule [Text]

The routing controls that are part of transactions that are evaluated to determine if a request to change a routing control state is allowed. For example, you might include three routing controls, one for each of three Amazon Web Services Regions.

assertionRule_ruleConfig :: Lens' AssertionRule RuleConfig

The criteria that you set for specific assertion controls (routing controls) that designate how many controls must be enabled as the result of a transaction. For example, if you have three assertion controls, you might specify atleast 2 for your rule configuration. This means that at least two assertion controls must be enabled, so that at least two Amazon Web Services Regions are enabled.

assertionRule_waitPeriodMs :: Lens' AssertionRule Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

assertionRule_name :: Lens' AssertionRule Text

Name of the assertion rule. You can use any non-white space character in the name.

AssertionRuleUpdate

data AssertionRuleUpdate

An update to an assertion rule. You can update the name or the evaluation period (wait period). If you don't specify one of the items to update, the item is unchanged.

See: newAssertionRuleUpdate smart constructor.

Constructors

AssertionRuleUpdate'

Fields

safetyRuleArn :: Text

The Amazon Resource Name (ARN) of the assertion rule.

waitPeriodMs :: Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

name :: Text

The name of the assertion rule. You can use any non-white space character in the name.

newAssertionRuleUpdate

Arguments

:: Text
$sel:safetyRuleArn:AssertionRuleUpdate'

-> Int
$sel:waitPeriodMs:AssertionRuleUpdate'

-> Text
$sel:name:AssertionRuleUpdate'

-> AssertionRuleUpdate

Create a value of AssertionRuleUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:safetyRuleArn:AssertionRuleUpdate', assertionRuleUpdate_safetyRuleArn - The Amazon Resource Name (ARN) of the assertion rule.

$sel:waitPeriodMs:AssertionRuleUpdate', assertionRuleUpdate_waitPeriodMs - An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

$sel:name:AssertionRuleUpdate', assertionRuleUpdate_name - The name of the assertion rule. You can use any non-white space character in the name.

assertionRuleUpdate_safetyRuleArn :: Lens' AssertionRuleUpdate Text

The Amazon Resource Name (ARN) of the assertion rule.

assertionRuleUpdate_waitPeriodMs :: Lens' AssertionRuleUpdate Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

assertionRuleUpdate_name :: Lens' AssertionRuleUpdate Text

The name of the assertion rule. You can use any non-white space character in the name.

Cluster

data Cluster

A cluster is a set of five consensus-forming Regional endpoints that represent the infrastructure that hosts your routing controls. Typically, you host together on one cluster all of the routing controls for your applications.

See: newCluster smart constructor.

Constructors

Cluster'

Fields

status :: Maybe Status

Deployment status of a resource. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

clusterArn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster.

name :: Maybe Text

The name of the cluster.

clusterEndpoints :: Maybe [ClusterEndpoint]

Endpoints for a cluster. Specify one of these endpoints when you want to set or retrieve a routing control state in the cluster.

To get or update the routing control state, see the Amazon Route 53 Application Recovery Controller Cluster (Data Plane) Actions.

newCluster :: Cluster

Create a value of Cluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Cluster', cluster_status - Deployment status of a resource. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

$sel:clusterArn:Cluster', cluster_clusterArn - The Amazon Resource Name (ARN) of the cluster.

$sel:name:Cluster', cluster_name - The name of the cluster.

$sel:clusterEndpoints:Cluster', cluster_clusterEndpoints - Endpoints for a cluster. Specify one of these endpoints when you want to set or retrieve a routing control state in the cluster.

To get or update the routing control state, see the Amazon Route 53 Application Recovery Controller Cluster (Data Plane) Actions.

cluster_status :: Lens' Cluster (Maybe Status)

Deployment status of a resource. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

cluster_clusterArn :: Lens' Cluster (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

cluster_name :: Lens' Cluster (Maybe Text)

The name of the cluster.

cluster_clusterEndpoints :: Lens' Cluster (Maybe [ClusterEndpoint])

Endpoints for a cluster. Specify one of these endpoints when you want to set or retrieve a routing control state in the cluster.

To get or update the routing control state, see the Amazon Route 53 Application Recovery Controller Cluster (Data Plane) Actions.

ClusterEndpoint

data ClusterEndpoint

A cluster endpoint. Specify an endpoint when you want to set or retrieve a routing control state in the cluster.

See: newClusterEndpoint smart constructor.

Constructors

ClusterEndpoint'

Fields

region :: Maybe Text

The Amazon Web Services Region for a cluster endpoint.

endpoint :: Maybe Text

A cluster endpoint. Specify an endpoint and Amazon Web Services Region when you want to set or retrieve a routing control state in the cluster.

To get or update the routing control state, see the Amazon Route 53 Application Recovery Controller Cluster (Data Plane) Actions.

newClusterEndpoint :: ClusterEndpoint

Create a value of ClusterEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:region:ClusterEndpoint', clusterEndpoint_region - The Amazon Web Services Region for a cluster endpoint.

$sel:endpoint:ClusterEndpoint', clusterEndpoint_endpoint - A cluster endpoint. Specify an endpoint and Amazon Web Services Region when you want to set or retrieve a routing control state in the cluster.

To get or update the routing control state, see the Amazon Route 53 Application Recovery Controller Cluster (Data Plane) Actions.

clusterEndpoint_region :: Lens' ClusterEndpoint (Maybe Text)

The Amazon Web Services Region for a cluster endpoint.

clusterEndpoint_endpoint :: Lens' ClusterEndpoint (Maybe Text)

A cluster endpoint. Specify an endpoint and Amazon Web Services Region when you want to set or retrieve a routing control state in the cluster.

To get or update the routing control state, see the Amazon Route 53 Application Recovery Controller Cluster (Data Plane) Actions.

ControlPanel

data ControlPanel

A control panel represents a group of routing controls that can be changed together in a single transaction.

See: newControlPanel smart constructor.

Constructors

ControlPanel'

Fields

status :: Maybe Status

The deployment status of control panel. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

controlPanelArn :: Maybe Text

The Amazon Resource Name (ARN) of the control panel.

clusterArn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster that includes the control panel.

routingControlCount :: Maybe Int

The number of routing controls in the control panel.

name :: Maybe Text

The name of the control panel. You can use any non-white space character in the name.

defaultControlPanel :: Maybe Bool

A flag that Amazon Route 53 Application Recovery Controller sets to true to designate the default control panel for a cluster. When you create a cluster, Amazon Route 53 Application Recovery Controller creates a control panel, and sets this flag for that control panel. If you create a control panel yourself, this flag is set to false.

newControlPanel :: ControlPanel

Create a value of ControlPanel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ControlPanel', controlPanel_status - The deployment status of control panel. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

$sel:controlPanelArn:ControlPanel', controlPanel_controlPanelArn - The Amazon Resource Name (ARN) of the control panel.

$sel:clusterArn:ControlPanel', controlPanel_clusterArn - The Amazon Resource Name (ARN) of the cluster that includes the control panel.

$sel:routingControlCount:ControlPanel', controlPanel_routingControlCount - The number of routing controls in the control panel.

$sel:name:ControlPanel', controlPanel_name - The name of the control panel. You can use any non-white space character in the name.

$sel:defaultControlPanel:ControlPanel', controlPanel_defaultControlPanel - A flag that Amazon Route 53 Application Recovery Controller sets to true to designate the default control panel for a cluster. When you create a cluster, Amazon Route 53 Application Recovery Controller creates a control panel, and sets this flag for that control panel. If you create a control panel yourself, this flag is set to false.

controlPanel_status :: Lens' ControlPanel (Maybe Status)

The deployment status of control panel. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

controlPanel_controlPanelArn :: Lens' ControlPanel (Maybe Text)

The Amazon Resource Name (ARN) of the control panel.

controlPanel_clusterArn :: Lens' ControlPanel (Maybe Text)

The Amazon Resource Name (ARN) of the cluster that includes the control panel.

controlPanel_routingControlCount :: Lens' ControlPanel (Maybe Int)

The number of routing controls in the control panel.

controlPanel_name :: Lens' ControlPanel (Maybe Text)

The name of the control panel. You can use any non-white space character in the name.

controlPanel_defaultControlPanel :: Lens' ControlPanel (Maybe Bool)

A flag that Amazon Route 53 Application Recovery Controller sets to true to designate the default control panel for a cluster. When you create a cluster, Amazon Route 53 Application Recovery Controller creates a control panel, and sets this flag for that control panel. If you create a control panel yourself, this flag is set to false.

GatingRule

data GatingRule

A gating rule verifies that a set of gating controls evaluates as true, based on a rule configuration that you specify. If the gating rule evaluates to true, Amazon Route 53 Application Recovery Controller allows a set of routing control state changes to run and complete against the set of target controls.

See: newGatingRule smart constructor.

Constructors

GatingRule'

Fields

status :: Status

The deployment status of a gating rule. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

targetControls :: [Text]

Routing controls that can only be set or unset if the specified RuleConfig evaluates to true for the specified GatingControls. For example, say you have three gating controls, one for each of three Amazon Web Services Regions. Now you specify ATLEAST 2 as your RuleConfig. With these settings, you can only change (set or unset) the routing controls that you have specified as TargetControls if that rule evaluates to true.

In other words, your ability to change the routing controls that you have specified as TargetControls is gated by the rule that you set for the routing controls in GatingControls.

controlPanelArn :: Text

The Amazon Resource Name (ARN) of the control panel.

safetyRuleArn :: Text

The Amazon Resource Name (ARN) of the gating rule.

gatingControls :: [Text]

The gating controls for the gating rule. That is, routing controls that are evaluated by the rule configuration that you specify.

ruleConfig :: RuleConfig

The criteria that you set for specific gating controls (routing controls) that designates how many controls must be enabled to allow you to change (set or unset) the target controls.

waitPeriodMs :: Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

name :: Text

The name for the gating rule.

newGatingRule

Arguments

:: Status
$sel:status:GatingRule'

-> Text
$sel:controlPanelArn:GatingRule'

-> Text
$sel:safetyRuleArn:GatingRule'

-> RuleConfig
$sel:ruleConfig:GatingRule'

-> Int
$sel:waitPeriodMs:GatingRule'

-> Text
$sel:name:GatingRule'

-> GatingRule

Create a value of GatingRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GatingRule', gatingRule_status - The deployment status of a gating rule. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

$sel:targetControls:GatingRule', gatingRule_targetControls - Routing controls that can only be set or unset if the specified RuleConfig evaluates to true for the specified GatingControls. For example, say you have three gating controls, one for each of three Amazon Web Services Regions. Now you specify ATLEAST 2 as your RuleConfig. With these settings, you can only change (set or unset) the routing controls that you have specified as TargetControls if that rule evaluates to true.

In other words, your ability to change the routing controls that you have specified as TargetControls is gated by the rule that you set for the routing controls in GatingControls.

$sel:controlPanelArn:GatingRule', gatingRule_controlPanelArn - The Amazon Resource Name (ARN) of the control panel.

$sel:safetyRuleArn:GatingRule', gatingRule_safetyRuleArn - The Amazon Resource Name (ARN) of the gating rule.

$sel:gatingControls:GatingRule', gatingRule_gatingControls - The gating controls for the gating rule. That is, routing controls that are evaluated by the rule configuration that you specify.

$sel:ruleConfig:GatingRule', gatingRule_ruleConfig - The criteria that you set for specific gating controls (routing controls) that designates how many controls must be enabled to allow you to change (set or unset) the target controls.

$sel:waitPeriodMs:GatingRule', gatingRule_waitPeriodMs - An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

$sel:name:GatingRule', gatingRule_name - The name for the gating rule.

gatingRule_status :: Lens' GatingRule Status

The deployment status of a gating rule. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

gatingRule_targetControls :: Lens' GatingRule [Text]

Routing controls that can only be set or unset if the specified RuleConfig evaluates to true for the specified GatingControls. For example, say you have three gating controls, one for each of three Amazon Web Services Regions. Now you specify ATLEAST 2 as your RuleConfig. With these settings, you can only change (set or unset) the routing controls that you have specified as TargetControls if that rule evaluates to true.

In other words, your ability to change the routing controls that you have specified as TargetControls is gated by the rule that you set for the routing controls in GatingControls.

gatingRule_controlPanelArn :: Lens' GatingRule Text

The Amazon Resource Name (ARN) of the control panel.

gatingRule_safetyRuleArn :: Lens' GatingRule Text

The Amazon Resource Name (ARN) of the gating rule.

gatingRule_gatingControls :: Lens' GatingRule [Text]

The gating controls for the gating rule. That is, routing controls that are evaluated by the rule configuration that you specify.

gatingRule_ruleConfig :: Lens' GatingRule RuleConfig

The criteria that you set for specific gating controls (routing controls) that designates how many controls must be enabled to allow you to change (set or unset) the target controls.

gatingRule_waitPeriodMs :: Lens' GatingRule Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

gatingRule_name :: Lens' GatingRule Text

The name for the gating rule.

GatingRuleUpdate

data GatingRuleUpdate

Update to a gating rule. You can update the name or the evaluation period (wait period). If you don't specify one of the items to update, the item is unchanged.

See: newGatingRuleUpdate smart constructor.

Constructors

GatingRuleUpdate'

Fields

safetyRuleArn :: Text

The Amazon Resource Name (ARN) of the gating rule.

waitPeriodMs :: Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

name :: Text

The name for the gating rule.

newGatingRuleUpdate

Arguments

:: Text
$sel:safetyRuleArn:GatingRuleUpdate'

-> Int
$sel:waitPeriodMs:GatingRuleUpdate'

-> Text
$sel:name:GatingRuleUpdate'

-> GatingRuleUpdate

Create a value of GatingRuleUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:safetyRuleArn:GatingRuleUpdate', gatingRuleUpdate_safetyRuleArn - The Amazon Resource Name (ARN) of the gating rule.

$sel:waitPeriodMs:GatingRuleUpdate', gatingRuleUpdate_waitPeriodMs - An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

$sel:name:GatingRuleUpdate', gatingRuleUpdate_name - The name for the gating rule.

gatingRuleUpdate_safetyRuleArn :: Lens' GatingRuleUpdate Text

The Amazon Resource Name (ARN) of the gating rule.

gatingRuleUpdate_waitPeriodMs :: Lens' GatingRuleUpdate Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

gatingRuleUpdate_name :: Lens' GatingRuleUpdate Text

The name for the gating rule.

NewAssertionRule

data NewAssertionRule

A new assertion rule for a control panel.

See: newNewAssertionRule smart constructor.

Constructors

NewAssertionRule'

Fields

controlPanelArn :: Text

The Amazon Resource Name (ARN) for the control panel.

assertedControls :: [Text]

The routing controls that are part of transactions that are evaluated to determine if a request to change a routing control state is allowed. For example, you might include three routing controls, one for each of three Amazon Web Services Regions.

ruleConfig :: RuleConfig

The criteria that you set for specific assertion controls (routing controls) that designate how many controls must be enabled as the result of a transaction. For example, if you have three assertion controls, you might specify atleast 2 for your rule configuration. This means that at least two assertion controls must be enabled, so that at least two Amazon Web Services Regions are enabled.

waitPeriodMs :: Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

name :: Text

The name of the assertion rule. You can use any non-white space character in the name.

newNewAssertionRule

Arguments

:: Text
$sel:controlPanelArn:NewAssertionRule'

-> RuleConfig
$sel:ruleConfig:NewAssertionRule'

-> Int
$sel:waitPeriodMs:NewAssertionRule'

-> Text
$sel:name:NewAssertionRule'

-> NewAssertionRule

Create a value of NewAssertionRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:controlPanelArn:NewAssertionRule', newAssertionRule_controlPanelArn - The Amazon Resource Name (ARN) for the control panel.

$sel:assertedControls:NewAssertionRule', newAssertionRule_assertedControls - The routing controls that are part of transactions that are evaluated to determine if a request to change a routing control state is allowed. For example, you might include three routing controls, one for each of three Amazon Web Services Regions.

$sel:ruleConfig:NewAssertionRule', newAssertionRule_ruleConfig - The criteria that you set for specific assertion controls (routing controls) that designate how many controls must be enabled as the result of a transaction. For example, if you have three assertion controls, you might specify atleast 2 for your rule configuration. This means that at least two assertion controls must be enabled, so that at least two Amazon Web Services Regions are enabled.

$sel:waitPeriodMs:NewAssertionRule', newAssertionRule_waitPeriodMs - An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

$sel:name:NewAssertionRule', newAssertionRule_name - The name of the assertion rule. You can use any non-white space character in the name.

newAssertionRule_controlPanelArn :: Lens' NewAssertionRule Text

The Amazon Resource Name (ARN) for the control panel.

newAssertionRule_assertedControls :: Lens' NewAssertionRule [Text]

The routing controls that are part of transactions that are evaluated to determine if a request to change a routing control state is allowed. For example, you might include three routing controls, one for each of three Amazon Web Services Regions.

newAssertionRule_ruleConfig :: Lens' NewAssertionRule RuleConfig

The criteria that you set for specific assertion controls (routing controls) that designate how many controls must be enabled as the result of a transaction. For example, if you have three assertion controls, you might specify atleast 2 for your rule configuration. This means that at least two assertion controls must be enabled, so that at least two Amazon Web Services Regions are enabled.

newAssertionRule_waitPeriodMs :: Lens' NewAssertionRule Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

newAssertionRule_name :: Lens' NewAssertionRule Text

The name of the assertion rule. You can use any non-white space character in the name.

NewGatingRule

data NewGatingRule

A new gating rule for a control panel.

See: newNewGatingRule smart constructor.

Constructors

NewGatingRule'

Fields

targetControls :: [Text]

Routing controls that can only be set or unset if the specified RuleConfig evaluates to true for the specified GatingControls. For example, say you have three gating controls, one for each of three Amazon Web Services Regions. Now you specify AtLeast 2 as your RuleConfig. With these settings, you can only change (set or unset) the routing controls that you have specified as TargetControls if that rule evaluates to true.

In other words, your ability to change the routing controls that you have specified as TargetControls is gated by the rule that you set for the routing controls in GatingControls.

controlPanelArn :: Text

The Amazon Resource Name (ARN) of the control panel.

gatingControls :: [Text]

The gating controls for the new gating rule. That is, routing controls that are evaluated by the rule configuration that you specify.

ruleConfig :: RuleConfig

The criteria that you set for specific gating controls (routing controls) that designates how many controls must be enabled to allow you to change (set or unset) the target controls.

waitPeriodMs :: Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

name :: Text

The name for the new gating rule.

newNewGatingRule

Arguments

:: Text
$sel:controlPanelArn:NewGatingRule'

-> RuleConfig
$sel:ruleConfig:NewGatingRule'

-> Int
$sel:waitPeriodMs:NewGatingRule'

-> Text
$sel:name:NewGatingRule'

-> NewGatingRule

Create a value of NewGatingRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetControls:NewGatingRule', newGatingRule_targetControls - Routing controls that can only be set or unset if the specified RuleConfig evaluates to true for the specified GatingControls. For example, say you have three gating controls, one for each of three Amazon Web Services Regions. Now you specify AtLeast 2 as your RuleConfig. With these settings, you can only change (set or unset) the routing controls that you have specified as TargetControls if that rule evaluates to true.

In other words, your ability to change the routing controls that you have specified as TargetControls is gated by the rule that you set for the routing controls in GatingControls.

$sel:controlPanelArn:NewGatingRule', newGatingRule_controlPanelArn - The Amazon Resource Name (ARN) of the control panel.

$sel:gatingControls:NewGatingRule', newGatingRule_gatingControls - The gating controls for the new gating rule. That is, routing controls that are evaluated by the rule configuration that you specify.

$sel:ruleConfig:NewGatingRule', newGatingRule_ruleConfig - The criteria that you set for specific gating controls (routing controls) that designates how many controls must be enabled to allow you to change (set or unset) the target controls.

$sel:waitPeriodMs:NewGatingRule', newGatingRule_waitPeriodMs - An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

$sel:name:NewGatingRule', newGatingRule_name - The name for the new gating rule.

newGatingRule_targetControls :: Lens' NewGatingRule [Text]

Routing controls that can only be set or unset if the specified RuleConfig evaluates to true for the specified GatingControls. For example, say you have three gating controls, one for each of three Amazon Web Services Regions. Now you specify AtLeast 2 as your RuleConfig. With these settings, you can only change (set or unset) the routing controls that you have specified as TargetControls if that rule evaluates to true.

In other words, your ability to change the routing controls that you have specified as TargetControls is gated by the rule that you set for the routing controls in GatingControls.

newGatingRule_controlPanelArn :: Lens' NewGatingRule Text

The Amazon Resource Name (ARN) of the control panel.

newGatingRule_gatingControls :: Lens' NewGatingRule [Text]

The gating controls for the new gating rule. That is, routing controls that are evaluated by the rule configuration that you specify.

newGatingRule_ruleConfig :: Lens' NewGatingRule RuleConfig

The criteria that you set for specific gating controls (routing controls) that designates how many controls must be enabled to allow you to change (set or unset) the target controls.

newGatingRule_waitPeriodMs :: Lens' NewGatingRule Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

newGatingRule_name :: Lens' NewGatingRule Text

The name for the new gating rule.

RoutingControl

data RoutingControl

A routing control has one of two states: ON and OFF. You can map the routing control state to the state of an Amazon Route 53 health check, which can be used to control traffic routing.

See: newRoutingControl smart constructor.

Constructors

RoutingControl'

Fields

status :: Maybe Status

The deployment status of a routing control. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

controlPanelArn :: Maybe Text

The Amazon Resource Name (ARN) of the control panel that includes the routing control.

name :: Maybe Text

The name of the routing control.

routingControlArn :: Maybe Text

The Amazon Resource Name (ARN) of the routing control.

newRoutingControl :: RoutingControl

Create a value of RoutingControl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:RoutingControl', routingControl_status - The deployment status of a routing control. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

$sel:controlPanelArn:RoutingControl', routingControl_controlPanelArn - The Amazon Resource Name (ARN) of the control panel that includes the routing control.

$sel:name:RoutingControl', routingControl_name - The name of the routing control.

$sel:routingControlArn:RoutingControl', routingControl_routingControlArn - The Amazon Resource Name (ARN) of the routing control.

routingControl_status :: Lens' RoutingControl (Maybe Status)

The deployment status of a routing control. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

routingControl_controlPanelArn :: Lens' RoutingControl (Maybe Text)

The Amazon Resource Name (ARN) of the control panel that includes the routing control.

routingControl_name :: Lens' RoutingControl (Maybe Text)

The name of the routing control.

routingControl_routingControlArn :: Lens' RoutingControl (Maybe Text)

The Amazon Resource Name (ARN) of the routing control.

Rule

data Rule

A safety rule. A safety rule can be an assertion rule or a gating rule.

See: newRule smart constructor.

Constructors

Rule'

Fields

gating :: Maybe GatingRule

A gating rule verifies that a set of gating controls evaluates as true, based on a rule configuration that you specify. If the gating rule evaluates to true, Amazon Route 53 Application Recovery Controller allows a set of routing control state changes to run and complete against the set of target controls.

assertion :: Maybe AssertionRule

An assertion rule enforces that, when a routing control state is changed, the criteria set by the rule configuration is met. Otherwise, the change to the routing control is not accepted.

newRule :: Rule

Create a value of Rule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gating:Rule', rule_gating - A gating rule verifies that a set of gating controls evaluates as true, based on a rule configuration that you specify. If the gating rule evaluates to true, Amazon Route 53 Application Recovery Controller allows a set of routing control state changes to run and complete against the set of target controls.

$sel:assertion:Rule', rule_assertion - An assertion rule enforces that, when a routing control state is changed, the criteria set by the rule configuration is met. Otherwise, the change to the routing control is not accepted.

rule_gating :: Lens' Rule (Maybe GatingRule)

A gating rule verifies that a set of gating controls evaluates as true, based on a rule configuration that you specify. If the gating rule evaluates to true, Amazon Route 53 Application Recovery Controller allows a set of routing control state changes to run and complete against the set of target controls.

rule_assertion :: Lens' Rule (Maybe AssertionRule)

An assertion rule enforces that, when a routing control state is changed, the criteria set by the rule configuration is met. Otherwise, the change to the routing control is not accepted.

RuleConfig

data RuleConfig

The rule configuration for an assertion rule. That is, the criteria that you set for specific assertion controls (routing controls) that specify how many controls must be enabled after a transaction completes.

See: newRuleConfig smart constructor.

Constructors

RuleConfig'

Fields

type' :: RuleType

A rule can be one of the following: ATLEAST, AND, or OR.

inverted :: Bool

Logical negation of the rule. If the rule would usually evaluate true, it's evaluated as false, and vice versa.

threshold :: Int

The value of N, when you specify an ATLEAST rule type. That is, Threshold is the number of controls that must be set when you specify an ATLEAST type.

newRuleConfig

Arguments

:: RuleType
$sel:type':RuleConfig'

-> Bool
$sel:inverted:RuleConfig'

-> Int
$sel:threshold:RuleConfig'

-> RuleConfig

Create a value of RuleConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':RuleConfig', ruleConfig_type - A rule can be one of the following: ATLEAST, AND, or OR.

$sel:inverted:RuleConfig', ruleConfig_inverted - Logical negation of the rule. If the rule would usually evaluate true, it's evaluated as false, and vice versa.

$sel:threshold:RuleConfig', ruleConfig_threshold - The value of N, when you specify an ATLEAST rule type. That is, Threshold is the number of controls that must be set when you specify an ATLEAST type.

ruleConfig_type :: Lens' RuleConfig RuleType

A rule can be one of the following: ATLEAST, AND, or OR.

ruleConfig_inverted :: Lens' RuleConfig Bool

Logical negation of the rule. If the rule would usually evaluate true, it's evaluated as false, and vice versa.

ruleConfig_threshold :: Lens' RuleConfig Int

The value of N, when you specify an ATLEAST rule type. That is, Threshold is the number of controls that must be set when you specify an ATLEAST type.