Network.AWS.Route53RecoveryControlConfig.Types.NewAssertionRule

Description

Documentation

data NewAssertionRule

A new assertion rule for a control panel.

See: newNewAssertionRule smart constructor.

Constructors

NewAssertionRule'

Fields

controlPanelArn :: Text

The Amazon Resource Name (ARN) for the control panel.

assertedControls :: [Text]

The routing controls that are part of transactions that are evaluated to determine if a request to change a routing control state is allowed. For example, you might include three routing controls, one for each of three Amazon Web Services Regions.

ruleConfig :: RuleConfig

The criteria that you set for specific assertion controls (routing controls) that designate how many controls must be enabled as the result of a transaction. For example, if you have three assertion controls, you might specify atleast 2 for your rule configuration. This means that at least two assertion controls must be enabled, so that at least two Amazon Web Services Regions are enabled.

waitPeriodMs :: Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

name :: Text

The name of the assertion rule. You can use any non-white space character in the name.

newNewAssertionRule

Arguments

:: Text
$sel:controlPanelArn:NewAssertionRule'

-> RuleConfig
$sel:ruleConfig:NewAssertionRule'

-> Int
$sel:waitPeriodMs:NewAssertionRule'

-> Text
$sel:name:NewAssertionRule'

-> NewAssertionRule

Create a value of NewAssertionRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:controlPanelArn:NewAssertionRule', newAssertionRule_controlPanelArn - The Amazon Resource Name (ARN) for the control panel.

$sel:assertedControls:NewAssertionRule', newAssertionRule_assertedControls - The routing controls that are part of transactions that are evaluated to determine if a request to change a routing control state is allowed. For example, you might include three routing controls, one for each of three Amazon Web Services Regions.

$sel:ruleConfig:NewAssertionRule', newAssertionRule_ruleConfig - The criteria that you set for specific assertion controls (routing controls) that designate how many controls must be enabled as the result of a transaction. For example, if you have three assertion controls, you might specify atleast 2 for your rule configuration. This means that at least two assertion controls must be enabled, so that at least two Amazon Web Services Regions are enabled.

$sel:waitPeriodMs:NewAssertionRule', newAssertionRule_waitPeriodMs - An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

$sel:name:NewAssertionRule', newAssertionRule_name - The name of the assertion rule. You can use any non-white space character in the name.

newAssertionRule_controlPanelArn :: Lens' NewAssertionRule Text

The Amazon Resource Name (ARN) for the control panel.

newAssertionRule_assertedControls :: Lens' NewAssertionRule [Text]

The routing controls that are part of transactions that are evaluated to determine if a request to change a routing control state is allowed. For example, you might include three routing controls, one for each of three Amazon Web Services Regions.

newAssertionRule_ruleConfig :: Lens' NewAssertionRule RuleConfig

The criteria that you set for specific assertion controls (routing controls) that designate how many controls must be enabled as the result of a transaction. For example, if you have three assertion controls, you might specify atleast 2 for your rule configuration. This means that at least two assertion controls must be enabled, so that at least two Amazon Web Services Regions are enabled.

newAssertionRule_waitPeriodMs :: Lens' NewAssertionRule Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

newAssertionRule_name :: Lens' NewAssertionRule Text

The name of the assertion rule. You can use any non-white space character in the name.