Network.AWS.Route53RecoveryControlConfig.Types.NewGatingRule

Description

Documentation

data NewGatingRule

A new gating rule for a control panel.

See: newNewGatingRule smart constructor.

Constructors

NewGatingRule'

Fields

targetControls :: [Text]

Routing controls that can only be set or unset if the specified RuleConfig evaluates to true for the specified GatingControls. For example, say you have three gating controls, one for each of three Amazon Web Services Regions. Now you specify AtLeast 2 as your RuleConfig. With these settings, you can only change (set or unset) the routing controls that you have specified as TargetControls if that rule evaluates to true.

In other words, your ability to change the routing controls that you have specified as TargetControls is gated by the rule that you set for the routing controls in GatingControls.

controlPanelArn :: Text

The Amazon Resource Name (ARN) of the control panel.

gatingControls :: [Text]

The gating controls for the new gating rule. That is, routing controls that are evaluated by the rule configuration that you specify.

ruleConfig :: RuleConfig

The criteria that you set for specific gating controls (routing controls) that designates how many controls must be enabled to allow you to change (set or unset) the target controls.

waitPeriodMs :: Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

name :: Text

The name for the new gating rule.

newNewGatingRule

Arguments

:: Text
$sel:controlPanelArn:NewGatingRule'

-> RuleConfig
$sel:ruleConfig:NewGatingRule'

-> Int
$sel:waitPeriodMs:NewGatingRule'

-> Text
$sel:name:NewGatingRule'

-> NewGatingRule

Create a value of NewGatingRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetControls:NewGatingRule', newGatingRule_targetControls - Routing controls that can only be set or unset if the specified RuleConfig evaluates to true for the specified GatingControls. For example, say you have three gating controls, one for each of three Amazon Web Services Regions. Now you specify AtLeast 2 as your RuleConfig. With these settings, you can only change (set or unset) the routing controls that you have specified as TargetControls if that rule evaluates to true.

In other words, your ability to change the routing controls that you have specified as TargetControls is gated by the rule that you set for the routing controls in GatingControls.

$sel:controlPanelArn:NewGatingRule', newGatingRule_controlPanelArn - The Amazon Resource Name (ARN) of the control panel.

$sel:gatingControls:NewGatingRule', newGatingRule_gatingControls - The gating controls for the new gating rule. That is, routing controls that are evaluated by the rule configuration that you specify.

$sel:ruleConfig:NewGatingRule', newGatingRule_ruleConfig - The criteria that you set for specific gating controls (routing controls) that designates how many controls must be enabled to allow you to change (set or unset) the target controls.

$sel:waitPeriodMs:NewGatingRule', newGatingRule_waitPeriodMs - An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

$sel:name:NewGatingRule', newGatingRule_name - The name for the new gating rule.

newGatingRule_targetControls :: Lens' NewGatingRule [Text]

Routing controls that can only be set or unset if the specified RuleConfig evaluates to true for the specified GatingControls. For example, say you have three gating controls, one for each of three Amazon Web Services Regions. Now you specify AtLeast 2 as your RuleConfig. With these settings, you can only change (set or unset) the routing controls that you have specified as TargetControls if that rule evaluates to true.

In other words, your ability to change the routing controls that you have specified as TargetControls is gated by the rule that you set for the routing controls in GatingControls.

newGatingRule_controlPanelArn :: Lens' NewGatingRule Text

The Amazon Resource Name (ARN) of the control panel.

newGatingRule_gatingControls :: Lens' NewGatingRule [Text]

The gating controls for the new gating rule. That is, routing controls that are evaluated by the rule configuration that you specify.

newGatingRule_ruleConfig :: Lens' NewGatingRule RuleConfig

The criteria that you set for specific gating controls (routing controls) that designates how many controls must be enabled to allow you to change (set or unset) the target controls.

newGatingRule_waitPeriodMs :: Lens' NewGatingRule Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

newGatingRule_name :: Lens' NewGatingRule Text

The name for the new gating rule.