Network.AWS.Route53RecoveryReadiness.GetArchitectureRecommendations

Description

Returns a collection of recommendations to improve resilliance and readiness check quality for a Recovery Group.

Creating a Request

data GetArchitectureRecommendations

See: newGetArchitectureRecommendations smart constructor.

Constructors

GetArchitectureRecommendations'

Fields

nextToken :: Maybe Text

A token that can be used to resume pagination from the end of the collection.

maxResults :: Maybe Natural

Upper bound on number of records to return.

recoveryGroupName :: Text

Name of RecoveryGroup (top level resource) to be analyzed.

newGetArchitectureRecommendations

Arguments

:: Text
$sel:recoveryGroupName:GetArchitectureRecommendations'

-> GetArchitectureRecommendations

Create a value of GetArchitectureRecommendations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetArchitectureRecommendations', getArchitectureRecommendations_nextToken - A token that can be used to resume pagination from the end of the collection.

$sel:maxResults:GetArchitectureRecommendations', getArchitectureRecommendations_maxResults - Upper bound on number of records to return.

$sel:recoveryGroupName:GetArchitectureRecommendations', getArchitectureRecommendations_recoveryGroupName - Name of RecoveryGroup (top level resource) to be analyzed.

Request Lenses

getArchitectureRecommendations_nextToken :: Lens' GetArchitectureRecommendations (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

getArchitectureRecommendations_maxResults :: Lens' GetArchitectureRecommendations (Maybe Natural)

Upper bound on number of records to return.

getArchitectureRecommendations_recoveryGroupName :: Lens' GetArchitectureRecommendations Text

Name of RecoveryGroup (top level resource) to be analyzed.

Destructuring the Response

data GetArchitectureRecommendationsResponse

See: newGetArchitectureRecommendationsResponse smart constructor.

Constructors

GetArchitectureRecommendationsResponse'

Fields

lastAuditTimestamp :: Maybe POSIX

The time a Recovery Group was last assessed for recommendations in UTC ISO-8601 format.

nextToken :: Maybe Text

A token that can be used to resume pagination from the end of the collection

recommendations :: Maybe [Recommendation]

A list of recommendations for the customer's application

httpStatus :: Int

The response's http status code.

newGetArchitectureRecommendationsResponse

Arguments

:: Int
$sel:httpStatus:GetArchitectureRecommendationsResponse'

-> GetArchitectureRecommendationsResponse

Create a value of GetArchitectureRecommendationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastAuditTimestamp:GetArchitectureRecommendationsResponse', getArchitectureRecommendationsResponse_lastAuditTimestamp - The time a Recovery Group was last assessed for recommendations in UTC ISO-8601 format.

$sel:nextToken:GetArchitectureRecommendations', getArchitectureRecommendationsResponse_nextToken - A token that can be used to resume pagination from the end of the collection

$sel:recommendations:GetArchitectureRecommendationsResponse', getArchitectureRecommendationsResponse_recommendations - A list of recommendations for the customer's application

$sel:httpStatus:GetArchitectureRecommendationsResponse', getArchitectureRecommendationsResponse_httpStatus - The response's http status code.

Response Lenses

getArchitectureRecommendationsResponse_lastAuditTimestamp :: Lens' GetArchitectureRecommendationsResponse (Maybe UTCTime)

The time a Recovery Group was last assessed for recommendations in UTC ISO-8601 format.

getArchitectureRecommendationsResponse_nextToken :: Lens' GetArchitectureRecommendationsResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection

getArchitectureRecommendationsResponse_recommendations :: Lens' GetArchitectureRecommendationsResponse (Maybe [Recommendation])

A list of recommendations for the customer's application

getArchitectureRecommendationsResponse_httpStatus :: Lens' GetArchitectureRecommendationsResponse Int

The response's http status code.