Network.AWS.Route53RecoveryReadiness.Lens

Description

Operations

CreateResourceSet

createResourceSet_tags :: Lens' CreateResourceSet (Maybe (HashMap Text Text))

Undocumented member.

createResourceSet_resourceSetType :: Lens' CreateResourceSet Text

AWS Resource type of the resources in the ResourceSet

createResourceSet_resourceSetName :: Lens' CreateResourceSet Text

The name of the ResourceSet to create

createResourceSet_resources :: Lens' CreateResourceSet [Resource]

A list of Resource objects

createResourceSetResponse_resourceSetName :: Lens' CreateResourceSetResponse (Maybe Text)

The name of the ResourceSet

createResourceSetResponse_resourceSetType :: Lens' CreateResourceSetResponse (Maybe Text)

AWS Resource Type of the resources in the ResourceSet

createResourceSetResponse_resources :: Lens' CreateResourceSetResponse (Maybe [Resource])

A list of Resource objects

createResourceSetResponse_resourceSetArn :: Lens' CreateResourceSetResponse (Maybe Text)

The arn for the ResourceSet

createResourceSetResponse_tags :: Lens' CreateResourceSetResponse (Maybe (HashMap Text Text))

Undocumented member.

createResourceSetResponse_httpStatus :: Lens' CreateResourceSetResponse Int

The response's http status code.

GetReadinessCheckStatus

getReadinessCheckStatus_nextToken :: Lens' GetReadinessCheckStatus (Maybe Text)

A token used to resume pagination from the end of a previous request.

getReadinessCheckStatus_maxResults :: Lens' GetReadinessCheckStatus (Maybe Natural)

Upper bound on number of records to return.

getReadinessCheckStatus_readinessCheckName :: Lens' GetReadinessCheckStatus Text

The ReadinessCheck to get

getReadinessCheckStatusResponse_readiness :: Lens' GetReadinessCheckStatusResponse (Maybe Readiness)

The readiness at rule level.

getReadinessCheckStatusResponse_resources :: Lens' GetReadinessCheckStatusResponse (Maybe [ResourceResult])

Summary of resources's readiness

getReadinessCheckStatusResponse_nextToken :: Lens' GetReadinessCheckStatusResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

getReadinessCheckStatusResponse_messages :: Lens' GetReadinessCheckStatusResponse (Maybe [Message])

Top level messages for readiness check status

getReadinessCheckStatusResponse_httpStatus :: Lens' GetReadinessCheckStatusResponse Int

The response's http status code.

GetCellReadinessSummary

getCellReadinessSummary_nextToken :: Lens' GetCellReadinessSummary (Maybe Text)

A token used to resume pagination from the end of a previous request.

getCellReadinessSummary_maxResults :: Lens' GetCellReadinessSummary (Maybe Natural)

Upper bound on number of records to return.

getCellReadinessSummary_cellName :: Lens' GetCellReadinessSummary Text

The name of the Cell

getCellReadinessSummaryResponse_readinessChecks :: Lens' GetCellReadinessSummaryResponse (Maybe [ReadinessCheckSummary])

Summaries for the ReadinessChecks making up the Cell

getCellReadinessSummaryResponse_readiness :: Lens' GetCellReadinessSummaryResponse (Maybe Readiness)

The readiness at Cell level.

getCellReadinessSummaryResponse_nextToken :: Lens' GetCellReadinessSummaryResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

getCellReadinessSummaryResponse_httpStatus :: Lens' GetCellReadinessSummaryResponse Int

The response's http status code.

UpdateCell

updateCell_cellName :: Lens' UpdateCell Text

The Cell to update

updateCell_cells :: Lens' UpdateCell [Text]

A list of Cell arns, completely replaces previous list

updateCellResponse_cells :: Lens' UpdateCellResponse (Maybe [Text])

A list of Cell arns

updateCellResponse_parentReadinessScopes :: Lens' UpdateCellResponse (Maybe [Text])

A list of Cell ARNs and/or RecoveryGroup ARNs

updateCellResponse_cellName :: Lens' UpdateCellResponse (Maybe Text)

The name of the Cell

updateCellResponse_cellArn :: Lens' UpdateCellResponse (Maybe Text)

The arn for the Cell

updateCellResponse_tags :: Lens' UpdateCellResponse (Maybe (HashMap Text Text))

Undocumented member.

updateCellResponse_httpStatus :: Lens' UpdateCellResponse Int

The response's http status code.

DeleteCell

deleteCell_cellName :: Lens' DeleteCell Text

The Cell to delete

UpdateReadinessCheck

updateReadinessCheck_readinessCheckName :: Lens' UpdateReadinessCheck Text

The ReadinessCheck to update

updateReadinessCheck_resourceSetName :: Lens' UpdateReadinessCheck Text

The name of the ResourceSet to check

updateReadinessCheckResponse_readinessCheckName :: Lens' UpdateReadinessCheckResponse (Maybe Text)

Name for a ReadinessCheck

updateReadinessCheckResponse_resourceSet :: Lens' UpdateReadinessCheckResponse (Maybe Text)

Name of the ResourceSet to be checked

updateReadinessCheckResponse_readinessCheckArn :: Lens' UpdateReadinessCheckResponse (Maybe Text)

Arn associated with ReadinessCheck

updateReadinessCheckResponse_tags :: Lens' UpdateReadinessCheckResponse (Maybe (HashMap Text Text))

Undocumented member.

updateReadinessCheckResponse_httpStatus :: Lens' UpdateReadinessCheckResponse Int

The response's http status code.

DeleteReadinessCheck

deleteReadinessCheck_readinessCheckName :: Lens' DeleteReadinessCheck Text

The ReadinessCheck to delete

ListCells

listCells_nextToken :: Lens' ListCells (Maybe Text)

A token used to resume pagination from the end of a previous request.

listCells_maxResults :: Lens' ListCells (Maybe Natural)

Upper bound on number of records to return.

listCellsResponse_cells :: Lens' ListCellsResponse (Maybe [CellOutput])

A list of Cells

listCellsResponse_nextToken :: Lens' ListCellsResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

listCellsResponse_httpStatus :: Lens' ListCellsResponse Int

The response's http status code.

ListReadinessChecks

listReadinessChecks_nextToken :: Lens' ListReadinessChecks (Maybe Text)

A token used to resume pagination from the end of a previous request.

listReadinessChecks_maxResults :: Lens' ListReadinessChecks (Maybe Natural)

Upper bound on number of records to return.

listReadinessChecksResponse_readinessChecks :: Lens' ListReadinessChecksResponse (Maybe [ReadinessCheckOutput])

A list of ReadinessCheck associated with the account

listReadinessChecksResponse_nextToken :: Lens' ListReadinessChecksResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

listReadinessChecksResponse_httpStatus :: Lens' ListReadinessChecksResponse Int

The response's http status code.

ListRules

listRules_resourceType :: Lens' ListRules (Maybe Text)

Filter parameter which specifies the rules to return given a resource type.

listRules_nextToken :: Lens' ListRules (Maybe Text)

A token used to resume pagination from the end of a previous request.

listRules_maxResults :: Lens' ListRules (Maybe Natural)

Upper bound on number of records to return.

listRulesResponse_rules :: Lens' ListRulesResponse (Maybe [ListRulesOutput])

A list of rules

listRulesResponse_nextToken :: Lens' ListRulesResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

listRulesResponse_httpStatus :: Lens' ListRulesResponse Int

The response's http status code.

CreateReadinessCheck

createReadinessCheck_tags :: Lens' CreateReadinessCheck (Maybe (HashMap Text Text))

Undocumented member.

createReadinessCheck_resourceSetName :: Lens' CreateReadinessCheck Text

The name of the ResourceSet to check

createReadinessCheck_readinessCheckName :: Lens' CreateReadinessCheck Text

The name of the ReadinessCheck to create

createReadinessCheckResponse_readinessCheckName :: Lens' CreateReadinessCheckResponse (Maybe Text)

Name for a ReadinessCheck

createReadinessCheckResponse_resourceSet :: Lens' CreateReadinessCheckResponse (Maybe Text)

Name of the ResourceSet to be checked

createReadinessCheckResponse_readinessCheckArn :: Lens' CreateReadinessCheckResponse (Maybe Text)

Arn associated with ReadinessCheck

createReadinessCheckResponse_tags :: Lens' CreateReadinessCheckResponse (Maybe (HashMap Text Text))

Undocumented member.

createReadinessCheckResponse_httpStatus :: Lens' CreateReadinessCheckResponse Int

The response's http status code.

CreateCell

createCell_cells :: Lens' CreateCell (Maybe [Text])

A list of Cell arns contained within this Cell (for use in nested Cells, e.g. regions within which AZs)

createCell_tags :: Lens' CreateCell (Maybe (HashMap Text Text))

Undocumented member.

createCell_cellName :: Lens' CreateCell Text

The name of the Cell to create

createCellResponse_cells :: Lens' CreateCellResponse (Maybe [Text])

A list of Cell arns

createCellResponse_parentReadinessScopes :: Lens' CreateCellResponse (Maybe [Text])

A list of Cell ARNs and/or RecoveryGroup ARNs

createCellResponse_cellName :: Lens' CreateCellResponse (Maybe Text)

The name of the Cell

createCellResponse_cellArn :: Lens' CreateCellResponse (Maybe Text)

The arn for the Cell

createCellResponse_tags :: Lens' CreateCellResponse (Maybe (HashMap Text Text))

Undocumented member.

createCellResponse_httpStatus :: Lens' CreateCellResponse Int

The response's http status code.

GetRecoveryGroup

getRecoveryGroup_recoveryGroupName :: Lens' GetRecoveryGroup Text

The RecoveryGroup to get

getRecoveryGroupResponse_cells :: Lens' GetRecoveryGroupResponse (Maybe [Text])

A list of Cell arns

getRecoveryGroupResponse_recoveryGroupName :: Lens' GetRecoveryGroupResponse (Maybe Text)

The name of the RecoveryGroup

getRecoveryGroupResponse_recoveryGroupArn :: Lens' GetRecoveryGroupResponse (Maybe Text)

The arn for the RecoveryGroup

getRecoveryGroupResponse_tags :: Lens' GetRecoveryGroupResponse (Maybe (HashMap Text Text))

Undocumented member.

getRecoveryGroupResponse_httpStatus :: Lens' GetRecoveryGroupResponse Int

The response's http status code.

ListRecoveryGroups

listRecoveryGroups_nextToken :: Lens' ListRecoveryGroups (Maybe Text)

A token used to resume pagination from the end of a previous request.

listRecoveryGroups_maxResults :: Lens' ListRecoveryGroups (Maybe Natural)

Upper bound on number of records to return.

listRecoveryGroupsResponse_recoveryGroups :: Lens' ListRecoveryGroupsResponse (Maybe [RecoveryGroupOutput])

A list of RecoveryGroups

listRecoveryGroupsResponse_nextToken :: Lens' ListRecoveryGroupsResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

listRecoveryGroupsResponse_httpStatus :: Lens' ListRecoveryGroupsResponse Int

The response's http status code.

ListCrossAccountAuthorizations

listCrossAccountAuthorizations_nextToken :: Lens' ListCrossAccountAuthorizations (Maybe Text)

A token used to resume pagination from the end of a previous request.

listCrossAccountAuthorizations_maxResults :: Lens' ListCrossAccountAuthorizations (Maybe Natural)

Upper bound on number of records to return.

listCrossAccountAuthorizationsResponse_crossAccountAuthorizations :: Lens' ListCrossAccountAuthorizationsResponse (Maybe [Text])

A list of CrossAccountAuthorizations

listCrossAccountAuthorizationsResponse_nextToken :: Lens' ListCrossAccountAuthorizationsResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

listCrossAccountAuthorizationsResponse_httpStatus :: Lens' ListCrossAccountAuthorizationsResponse Int

The response's http status code.

GetCell

getCell_cellName :: Lens' GetCell Text

The Cell to get

getCellResponse_cells :: Lens' GetCellResponse (Maybe [Text])

A list of Cell arns

getCellResponse_parentReadinessScopes :: Lens' GetCellResponse (Maybe [Text])

A list of Cell ARNs and/or RecoveryGroup ARNs

getCellResponse_cellName :: Lens' GetCellResponse (Maybe Text)

The name of the Cell

getCellResponse_cellArn :: Lens' GetCellResponse (Maybe Text)

The arn for the Cell

getCellResponse_tags :: Lens' GetCellResponse (Maybe (HashMap Text Text))

Undocumented member.

getCellResponse_httpStatus :: Lens' GetCellResponse Int

The response's http status code.

CreateCrossAccountAuthorization

createCrossAccountAuthorization_crossAccountAuthorization :: Lens' CreateCrossAccountAuthorization Text

The cross account authorization

createCrossAccountAuthorizationResponse_crossAccountAuthorization :: Lens' CreateCrossAccountAuthorizationResponse (Maybe Text)

The cross account authorization

createCrossAccountAuthorizationResponse_httpStatus :: Lens' CreateCrossAccountAuthorizationResponse Int

The response's http status code.

CreateRecoveryGroup

createRecoveryGroup_cells :: Lens' CreateRecoveryGroup (Maybe [Text])

A list of Cell arns

createRecoveryGroup_tags :: Lens' CreateRecoveryGroup (Maybe (HashMap Text Text))

Undocumented member.

createRecoveryGroup_recoveryGroupName :: Lens' CreateRecoveryGroup Text

The name of the RecoveryGroup to create

createRecoveryGroupResponse_cells :: Lens' CreateRecoveryGroupResponse (Maybe [Text])

A list of Cell arns

createRecoveryGroupResponse_recoveryGroupName :: Lens' CreateRecoveryGroupResponse (Maybe Text)

The name of the RecoveryGroup

createRecoveryGroupResponse_recoveryGroupArn :: Lens' CreateRecoveryGroupResponse (Maybe Text)

The arn for the RecoveryGroup

createRecoveryGroupResponse_tags :: Lens' CreateRecoveryGroupResponse (Maybe (HashMap Text Text))

Undocumented member.

createRecoveryGroupResponse_httpStatus :: Lens' CreateRecoveryGroupResponse Int

The response's http status code.

GetReadinessCheck

getReadinessCheck_readinessCheckName :: Lens' GetReadinessCheck Text

The ReadinessCheck to get

getReadinessCheckResponse_readinessCheckName :: Lens' GetReadinessCheckResponse (Maybe Text)

Name for a ReadinessCheck

getReadinessCheckResponse_resourceSet :: Lens' GetReadinessCheckResponse (Maybe Text)

Name of the ResourceSet to be checked

getReadinessCheckResponse_readinessCheckArn :: Lens' GetReadinessCheckResponse (Maybe Text)

Arn associated with ReadinessCheck

getReadinessCheckResponse_tags :: Lens' GetReadinessCheckResponse (Maybe (HashMap Text Text))

Undocumented member.

getReadinessCheckResponse_httpStatus :: Lens' GetReadinessCheckResponse Int

The response's http status code.

GetReadinessCheckResourceStatus

getReadinessCheckResourceStatus_nextToken :: Lens' GetReadinessCheckResourceStatus (Maybe Text)

A token used to resume pagination from the end of a previous request.

getReadinessCheckResourceStatus_maxResults :: Lens' GetReadinessCheckResourceStatus (Maybe Natural)

Upper bound on number of records to return.

getReadinessCheckResourceStatus_readinessCheckName :: Lens' GetReadinessCheckResourceStatus Text

The ReadinessCheck to get

getReadinessCheckResourceStatus_resourceIdentifier :: Lens' GetReadinessCheckResourceStatus Text

The resource ARN or component Id to get

getReadinessCheckResourceStatusResponse_rules :: Lens' GetReadinessCheckResourceStatusResponse (Maybe [RuleResult])

Details of the rules's results

getReadinessCheckResourceStatusResponse_readiness :: Lens' GetReadinessCheckResourceStatusResponse (Maybe Readiness)

The readiness at rule level.

getReadinessCheckResourceStatusResponse_nextToken :: Lens' GetReadinessCheckResourceStatusResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

getReadinessCheckResourceStatusResponse_httpStatus :: Lens' GetReadinessCheckResourceStatusResponse Int

The response's http status code.

ListResourceSets

listResourceSets_nextToken :: Lens' ListResourceSets (Maybe Text)

A token used to resume pagination from the end of a previous request.

listResourceSets_maxResults :: Lens' ListResourceSets (Maybe Natural)

Upper bound on number of records to return.

listResourceSetsResponse_resourceSets :: Lens' ListResourceSetsResponse (Maybe [ResourceSetOutput])

A list of ResourceSets associated with the account

listResourceSetsResponse_nextToken :: Lens' ListResourceSetsResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

listResourceSetsResponse_httpStatus :: Lens' ListResourceSetsResponse Int

The response's http status code.

GetArchitectureRecommendations

getArchitectureRecommendations_nextToken :: Lens' GetArchitectureRecommendations (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

getArchitectureRecommendations_maxResults :: Lens' GetArchitectureRecommendations (Maybe Natural)

Upper bound on number of records to return.

getArchitectureRecommendations_recoveryGroupName :: Lens' GetArchitectureRecommendations Text

Name of RecoveryGroup (top level resource) to be analyzed.

getArchitectureRecommendationsResponse_lastAuditTimestamp :: Lens' GetArchitectureRecommendationsResponse (Maybe UTCTime)

The time a Recovery Group was last assessed for recommendations in UTC ISO-8601 format.

getArchitectureRecommendationsResponse_nextToken :: Lens' GetArchitectureRecommendationsResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection

getArchitectureRecommendationsResponse_recommendations :: Lens' GetArchitectureRecommendationsResponse (Maybe [Recommendation])

A list of recommendations for the customer's application

getArchitectureRecommendationsResponse_httpStatus :: Lens' GetArchitectureRecommendationsResponse Int

The response's http status code.

DeleteCrossAccountAuthorization

deleteCrossAccountAuthorization_crossAccountAuthorization :: Lens' DeleteCrossAccountAuthorization Text

The cross account authorization

deleteCrossAccountAuthorizationResponse_httpStatus :: Lens' DeleteCrossAccountAuthorizationResponse Int

The response's http status code.

DeleteRecoveryGroup

deleteRecoveryGroup_recoveryGroupName :: Lens' DeleteRecoveryGroup Text

The RecoveryGroup to delete

UpdateRecoveryGroup

updateRecoveryGroup_recoveryGroupName :: Lens' UpdateRecoveryGroup Text

The RecoveryGroup to update

updateRecoveryGroup_cells :: Lens' UpdateRecoveryGroup [Text]

A list of Cell arns, completely replaces previous list

updateRecoveryGroupResponse_cells :: Lens' UpdateRecoveryGroupResponse (Maybe [Text])

A list of Cell arns

updateRecoveryGroupResponse_recoveryGroupName :: Lens' UpdateRecoveryGroupResponse (Maybe Text)

The name of the RecoveryGroup

updateRecoveryGroupResponse_recoveryGroupArn :: Lens' UpdateRecoveryGroupResponse (Maybe Text)

The arn for the RecoveryGroup

updateRecoveryGroupResponse_tags :: Lens' UpdateRecoveryGroupResponse (Maybe (HashMap Text Text))

Undocumented member.

updateRecoveryGroupResponse_httpStatus :: Lens' UpdateRecoveryGroupResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) for the resource. You can get this from the response to any request to the resource.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

Undocumented member.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_tagKeys :: Lens' UntagResource [Text]

A comma-separated list of the tag keys to remove from the resource.

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) for the resource. You can get this from the response to any request to the resource.

GetRecoveryGroupReadinessSummary

getRecoveryGroupReadinessSummary_nextToken :: Lens' GetRecoveryGroupReadinessSummary (Maybe Text)

A token used to resume pagination from the end of a previous request.

getRecoveryGroupReadinessSummary_maxResults :: Lens' GetRecoveryGroupReadinessSummary (Maybe Natural)

Upper bound on number of records to return.

getRecoveryGroupReadinessSummary_recoveryGroupName :: Lens' GetRecoveryGroupReadinessSummary Text

The name of the RecoveryGroup

getRecoveryGroupReadinessSummaryResponse_readinessChecks :: Lens' GetRecoveryGroupReadinessSummaryResponse (Maybe [ReadinessCheckSummary])

Summaries for the ReadinessChecks making up the RecoveryGroup

getRecoveryGroupReadinessSummaryResponse_readiness :: Lens' GetRecoveryGroupReadinessSummaryResponse (Maybe Readiness)

The readiness at RecoveryGroup level.

getRecoveryGroupReadinessSummaryResponse_nextToken :: Lens' GetRecoveryGroupReadinessSummaryResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

getRecoveryGroupReadinessSummaryResponse_httpStatus :: Lens' GetRecoveryGroupReadinessSummaryResponse Int

The response's http status code.

GetResourceSet

getResourceSet_resourceSetName :: Lens' GetResourceSet Text

The ResourceSet to get

getResourceSetResponse_resourceSetName :: Lens' GetResourceSetResponse (Maybe Text)

The name of the ResourceSet

getResourceSetResponse_resourceSetType :: Lens' GetResourceSetResponse (Maybe Text)

AWS Resource Type of the resources in the ResourceSet

getResourceSetResponse_resources :: Lens' GetResourceSetResponse (Maybe [Resource])

A list of Resource objects

getResourceSetResponse_resourceSetArn :: Lens' GetResourceSetResponse (Maybe Text)

The arn for the ResourceSet

getResourceSetResponse_tags :: Lens' GetResourceSetResponse (Maybe (HashMap Text Text))

Undocumented member.

getResourceSetResponse_httpStatus :: Lens' GetResourceSetResponse Int

The response's http status code.

ListTagsForResources

listTagsForResources_resourceArn :: Lens' ListTagsForResources Text

The Amazon Resource Name (ARN) for the resource. You can get this from the response to any request to the resource.

listTagsForResourcesResponse_tags :: Lens' ListTagsForResourcesResponse (Maybe (HashMap Text Text))

Undocumented member.

listTagsForResourcesResponse_httpStatus :: Lens' ListTagsForResourcesResponse Int

The response's http status code.

UpdateResourceSet

updateResourceSet_resourceSetName :: Lens' UpdateResourceSet Text

The ResourceSet to update

updateResourceSet_resourceSetType :: Lens' UpdateResourceSet Text

AWS Resource Type of the resources in the ResourceSet

updateResourceSet_resources :: Lens' UpdateResourceSet [Resource]

A list of Resource objects

updateResourceSetResponse_resourceSetName :: Lens' UpdateResourceSetResponse (Maybe Text)

The name of the ResourceSet

updateResourceSetResponse_resourceSetType :: Lens' UpdateResourceSetResponse (Maybe Text)

AWS Resource Type of the resources in the ResourceSet

updateResourceSetResponse_resources :: Lens' UpdateResourceSetResponse (Maybe [Resource])

A list of Resource objects

updateResourceSetResponse_resourceSetArn :: Lens' UpdateResourceSetResponse (Maybe Text)

The arn for the ResourceSet

updateResourceSetResponse_tags :: Lens' UpdateResourceSetResponse (Maybe (HashMap Text Text))

Undocumented member.

updateResourceSetResponse_httpStatus :: Lens' UpdateResourceSetResponse Int

The response's http status code.

DeleteResourceSet

deleteResourceSet_resourceSetName :: Lens' DeleteResourceSet Text

The ResourceSet to delete

Types

CellOutput

cellOutput_tags :: Lens' CellOutput (Maybe (HashMap Text Text))

Undocumented member.

cellOutput_parentReadinessScopes :: Lens' CellOutput [Text]

A list of Cell ARNs and/or RecoveryGroup ARNs

cellOutput_cellArn :: Lens' CellOutput Text

The arn for the Cell

cellOutput_cellName :: Lens' CellOutput Text

The name of the Cell

cellOutput_cells :: Lens' CellOutput [Text]

A list of Cell arns

DNSTargetResource

dNSTargetResource_hostedZoneArn :: Lens' DNSTargetResource (Maybe Text)

The Hosted Zone ARN that contains the DNS record with the provided name of target resource.

dNSTargetResource_recordType :: Lens' DNSTargetResource (Maybe Text)

The Type of DNS Record of target resource

dNSTargetResource_targetResource :: Lens' DNSTargetResource (Maybe TargetResource)

Undocumented member.

dNSTargetResource_domainName :: Lens' DNSTargetResource (Maybe Text)

The DNS Name that acts as ingress point to a portion of application

dNSTargetResource_recordSetId :: Lens' DNSTargetResource (Maybe Text)

The R53 Set Id to uniquely identify a record given a Name and a Type

ListRulesOutput

listRulesOutput_ruleDescription :: Lens' ListRulesOutput Text

A description of the rule

listRulesOutput_ruleId :: Lens' ListRulesOutput Text

The Rule's ID.

listRulesOutput_resourceType :: Lens' ListRulesOutput Text

The resource type the rule applies to.

Message

message_messageText :: Lens' Message (Maybe Text)

The text of a readiness check message

NLBResource

nLBResource_arn :: Lens' NLBResource (Maybe Text)

An NLB resource arn

R53ResourceRecord

r53ResourceRecord_domainName :: Lens' R53ResourceRecord (Maybe Text)

The DNS target name

r53ResourceRecord_recordSetId :: Lens' R53ResourceRecord (Maybe Text)

The Resource Record set id

ReadinessCheckOutput

readinessCheckOutput_readinessCheckName :: Lens' ReadinessCheckOutput (Maybe Text)

Name for a ReadinessCheck

readinessCheckOutput_tags :: Lens' ReadinessCheckOutput (Maybe (HashMap Text Text))

Undocumented member.

readinessCheckOutput_readinessCheckArn :: Lens' ReadinessCheckOutput Text

Arn associated with ReadinessCheck

readinessCheckOutput_resourceSet :: Lens' ReadinessCheckOutput Text

Name of the ResourceSet to be checked

ReadinessCheckSummary

readinessCheckSummary_readiness :: Lens' ReadinessCheckSummary (Maybe Readiness)

The readiness of this ReadinessCheck

readinessCheckSummary_readinessCheckName :: Lens' ReadinessCheckSummary (Maybe Text)

The name of a ReadinessCheck which is part of the given RecoveryGroup or Cell

Recommendation

recommendation_recommendationText :: Lens' Recommendation Text

Guidance text for recommendation

RecoveryGroupOutput

recoveryGroupOutput_tags :: Lens' RecoveryGroupOutput (Maybe (HashMap Text Text))

Undocumented member.

recoveryGroupOutput_recoveryGroupArn :: Lens' RecoveryGroupOutput Text

The arn for the RecoveryGroup

recoveryGroupOutput_recoveryGroupName :: Lens' RecoveryGroupOutput Text

The name of the RecoveryGroup

recoveryGroupOutput_cells :: Lens' RecoveryGroupOutput [Text]

A list of Cell arns

Resource

resource_readinessScopes :: Lens' Resource (Maybe [Text])

A list of RecoveryGroup ARNs and/or Cell ARNs that this resource is contained within.

resource_resourceArn :: Lens' Resource (Maybe Text)

The ARN of the AWS resource, can be skipped if dnsTargetResource is used

resource_componentId :: Lens' Resource (Maybe Text)

The component id of the resource, generated by the service when dnsTargetResource is used

resource_dnsTargetResource :: Lens' Resource (Maybe DNSTargetResource)

Undocumented member.

ResourceResult

resourceResult_resourceArn :: Lens' ResourceResult (Maybe Text)

The ARN of the resource

resourceResult_componentId :: Lens' ResourceResult (Maybe Text)

The component id of the resource

resourceResult_readiness :: Lens' ResourceResult Readiness

The readiness of the resource.

resourceResult_lastCheckedTimestamp :: Lens' ResourceResult UTCTime

The time the resource was last checked for readiness, in ISO-8601 format, UTC.

ResourceSetOutput

resourceSetOutput_tags :: Lens' ResourceSetOutput (Maybe (HashMap Text Text))

Undocumented member.

resourceSetOutput_resourceSetType :: Lens' ResourceSetOutput Text

AWS Resource Type of the resources in the ResourceSet

resourceSetOutput_resourceSetName :: Lens' ResourceSetOutput Text

The name of the ResourceSet

resourceSetOutput_resourceSetArn :: Lens' ResourceSetOutput Text

The arn for the ResourceSet

resourceSetOutput_resources :: Lens' ResourceSetOutput [Resource]

A list of Resource objects

RuleResult

ruleResult_messages :: Lens' RuleResult [Message]

Details about the resource's readiness

ruleResult_readiness :: Lens' RuleResult Readiness

The readiness at rule level.

ruleResult_ruleId :: Lens' RuleResult Text

The identifier of the rule.

ruleResult_lastCheckedTimestamp :: Lens' RuleResult UTCTime

The time the resource was last checked for readiness, in ISO-8601 format, UTC.

TargetResource

targetResource_r53Resource :: Lens' TargetResource (Maybe R53ResourceRecord)

Undocumented member.

targetResource_nLBResource :: Lens' TargetResource (Maybe NLBResource)

Undocumented member.