Network.AWS.Route53RecoveryReadiness.ListCrossAccountAuthorizations

Description

Returns a collection of cross account readiness authorizations.

This operation returns paginated results.

Creating a Request

data ListCrossAccountAuthorizations

See: newListCrossAccountAuthorizations smart constructor.

Constructors

ListCrossAccountAuthorizations'

Fields

nextToken :: Maybe Text

A token used to resume pagination from the end of a previous request.

maxResults :: Maybe Natural

Upper bound on number of records to return.

newListCrossAccountAuthorizations :: ListCrossAccountAuthorizations

Create a value of ListCrossAccountAuthorizations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListCrossAccountAuthorizations', listCrossAccountAuthorizations_nextToken - A token used to resume pagination from the end of a previous request.

$sel:maxResults:ListCrossAccountAuthorizations', listCrossAccountAuthorizations_maxResults - Upper bound on number of records to return.

Request Lenses

listCrossAccountAuthorizations_nextToken :: Lens' ListCrossAccountAuthorizations (Maybe Text)

A token used to resume pagination from the end of a previous request.

listCrossAccountAuthorizations_maxResults :: Lens' ListCrossAccountAuthorizations (Maybe Natural)

Upper bound on number of records to return.

Destructuring the Response

data ListCrossAccountAuthorizationsResponse

See: newListCrossAccountAuthorizationsResponse smart constructor.

Constructors

ListCrossAccountAuthorizationsResponse'

Fields

crossAccountAuthorizations :: Maybe [Text]

A list of CrossAccountAuthorizations

nextToken :: Maybe Text

A token that can be used to resume pagination from the end of the collection.

httpStatus :: Int

The response's http status code.

newListCrossAccountAuthorizationsResponse

Arguments

:: Int
$sel:httpStatus:ListCrossAccountAuthorizationsResponse'

-> ListCrossAccountAuthorizationsResponse

Create a value of ListCrossAccountAuthorizationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:crossAccountAuthorizations:ListCrossAccountAuthorizationsResponse', listCrossAccountAuthorizationsResponse_crossAccountAuthorizations - A list of CrossAccountAuthorizations

$sel:nextToken:ListCrossAccountAuthorizations', listCrossAccountAuthorizationsResponse_nextToken - A token that can be used to resume pagination from the end of the collection.

$sel:httpStatus:ListCrossAccountAuthorizationsResponse', listCrossAccountAuthorizationsResponse_httpStatus - The response's http status code.

Response Lenses

listCrossAccountAuthorizationsResponse_crossAccountAuthorizations :: Lens' ListCrossAccountAuthorizationsResponse (Maybe [Text])

A list of CrossAccountAuthorizations

listCrossAccountAuthorizationsResponse_nextToken :: Lens' ListCrossAccountAuthorizationsResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

listCrossAccountAuthorizationsResponse_httpStatus :: Lens' ListCrossAccountAuthorizationsResponse Int

The response's http status code.