Network.AWS.Route53RecoveryReadiness.Types

Description

Service Configuration

defaultService :: Service

API version 2019-12-02 of the Amazon Route53 Recovery Readiness SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by an AWS service.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

User does not have sufficient access to perform this action.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Updating or deleting a resource can cause an inconsistent state.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

Request was denied due to request throttling.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

An unexpected error occurred.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The requested resource does not exist.

Readiness

newtype Readiness

The readiness of an entire ReadinessCheck or an individual resource ARN.

Constructors

Readiness'

Fields

fromReadiness :: Text

Bundled Patterns

pattern Readiness_NOT_AUTHORIZED :: Readiness

pattern Readiness_NOT_READY :: Readiness

pattern Readiness_READY :: Readiness

pattern Readiness_UNKNOWN :: Readiness

CellOutput

data CellOutput

A Cell and its properties

See: newCellOutput smart constructor.

Constructors

CellOutput'

Fields

tags :: Maybe (HashMap Text Text)

parentReadinessScopes :: [Text]

A list of Cell ARNs and/or RecoveryGroup ARNs

cellArn :: Text

The arn for the Cell

cellName :: Text

The name of the Cell

cells :: [Text]

A list of Cell arns

newCellOutput

Arguments

:: Text
$sel:cellArn:CellOutput'

-> Text
$sel:cellName:CellOutput'

-> CellOutput

Create a value of CellOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CellOutput', cellOutput_tags - Undocumented member.

$sel:parentReadinessScopes:CellOutput', cellOutput_parentReadinessScopes - A list of Cell ARNs and/or RecoveryGroup ARNs

$sel:cellArn:CellOutput', cellOutput_cellArn - The arn for the Cell

$sel:cellName:CellOutput', cellOutput_cellName - The name of the Cell

$sel:cells:CellOutput', cellOutput_cells - A list of Cell arns

cellOutput_tags :: Lens' CellOutput (Maybe (HashMap Text Text))

Undocumented member.

cellOutput_parentReadinessScopes :: Lens' CellOutput [Text]

A list of Cell ARNs and/or RecoveryGroup ARNs

cellOutput_cellArn :: Lens' CellOutput Text

The arn for the Cell

cellOutput_cellName :: Lens' CellOutput Text

The name of the Cell

cellOutput_cells :: Lens' CellOutput [Text]

A list of Cell arns

DNSTargetResource

data DNSTargetResource

A component for DNS/Routing Control Readiness Checks

See: newDNSTargetResource smart constructor.

Constructors

DNSTargetResource'

Fields

hostedZoneArn :: Maybe Text

The Hosted Zone ARN that contains the DNS record with the provided name of target resource.

recordType :: Maybe Text

The Type of DNS Record of target resource

targetResource :: Maybe TargetResource

domainName :: Maybe Text

The DNS Name that acts as ingress point to a portion of application

recordSetId :: Maybe Text

The R53 Set Id to uniquely identify a record given a Name and a Type

newDNSTargetResource :: DNSTargetResource

Create a value of DNSTargetResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostedZoneArn:DNSTargetResource', dNSTargetResource_hostedZoneArn - The Hosted Zone ARN that contains the DNS record with the provided name of target resource.

$sel:recordType:DNSTargetResource', dNSTargetResource_recordType - The Type of DNS Record of target resource

$sel:targetResource:DNSTargetResource', dNSTargetResource_targetResource - Undocumented member.

$sel:domainName:DNSTargetResource', dNSTargetResource_domainName - The DNS Name that acts as ingress point to a portion of application

$sel:recordSetId:DNSTargetResource', dNSTargetResource_recordSetId - The R53 Set Id to uniquely identify a record given a Name and a Type

dNSTargetResource_hostedZoneArn :: Lens' DNSTargetResource (Maybe Text)

The Hosted Zone ARN that contains the DNS record with the provided name of target resource.

dNSTargetResource_recordType :: Lens' DNSTargetResource (Maybe Text)

The Type of DNS Record of target resource

dNSTargetResource_targetResource :: Lens' DNSTargetResource (Maybe TargetResource)

Undocumented member.

dNSTargetResource_domainName :: Lens' DNSTargetResource (Maybe Text)

The DNS Name that acts as ingress point to a portion of application

dNSTargetResource_recordSetId :: Lens' DNSTargetResource (Maybe Text)

The R53 Set Id to uniquely identify a record given a Name and a Type

ListRulesOutput

data ListRulesOutput

A collection of rules used in a readiness check

See: newListRulesOutput smart constructor.

Constructors

ListRulesOutput'

Fields

ruleDescription :: Text

A description of the rule

ruleId :: Text

The Rule's ID.

resourceType :: Text

The resource type the rule applies to.

newListRulesOutput

Arguments

:: Text
$sel:ruleDescription:ListRulesOutput'

-> Text
$sel:ruleId:ListRulesOutput'

-> Text
$sel:resourceType:ListRulesOutput'

-> ListRulesOutput

Create a value of ListRulesOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleDescription:ListRulesOutput', listRulesOutput_ruleDescription - A description of the rule

$sel:ruleId:ListRulesOutput', listRulesOutput_ruleId - The Rule's ID.

$sel:resourceType:ListRulesOutput', listRulesOutput_resourceType - The resource type the rule applies to.

listRulesOutput_ruleDescription :: Lens' ListRulesOutput Text

A description of the rule

listRulesOutput_ruleId :: Lens' ListRulesOutput Text

The Rule's ID.

listRulesOutput_resourceType :: Lens' ListRulesOutput Text

The resource type the rule applies to.

Message

data Message

Information relating to readiness check status

See: newMessage smart constructor.

Constructors

Message'

Fields

messageText :: Maybe Text

The text of a readiness check message

newMessage :: Message

Create a value of Message with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:messageText:Message', message_messageText - The text of a readiness check message

message_messageText :: Lens' Message (Maybe Text)

The text of a readiness check message

NLBResource

data NLBResource

The NLB resource a DNS Target Resource points to

See: newNLBResource smart constructor.

Constructors

NLBResource'

Fields

arn :: Maybe Text

An NLB resource arn

newNLBResource :: NLBResource

Create a value of NLBResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:NLBResource', nLBResource_arn - An NLB resource arn

nLBResource_arn :: Lens' NLBResource (Maybe Text)

An NLB resource arn

R53ResourceRecord

data R53ResourceRecord

The Route 53 resource a DNS Target Resource record points to

See: newR53ResourceRecord smart constructor.

Constructors

R53ResourceRecord'

Fields

domainName :: Maybe Text

The DNS target name

recordSetId :: Maybe Text

The Resource Record set id

newR53ResourceRecord :: R53ResourceRecord

Create a value of R53ResourceRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:R53ResourceRecord', r53ResourceRecord_domainName - The DNS target name

$sel:recordSetId:R53ResourceRecord', r53ResourceRecord_recordSetId - The Resource Record set id

r53ResourceRecord_domainName :: Lens' R53ResourceRecord (Maybe Text)

The DNS target name

r53ResourceRecord_recordSetId :: Lens' R53ResourceRecord (Maybe Text)

The Resource Record set id

ReadinessCheckOutput

data ReadinessCheckOutput

A resource used for checking the readiness of a Resource Set

See: newReadinessCheckOutput smart constructor.

Constructors

ReadinessCheckOutput'

Fields

readinessCheckName :: Maybe Text

Name for a ReadinessCheck

tags :: Maybe (HashMap Text Text)

readinessCheckArn :: Text

Arn associated with ReadinessCheck

resourceSet :: Text

Name of the ResourceSet to be checked

newReadinessCheckOutput

Arguments

:: Text
$sel:readinessCheckArn:ReadinessCheckOutput'

-> Text
$sel:resourceSet:ReadinessCheckOutput'

-> ReadinessCheckOutput

Create a value of ReadinessCheckOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:readinessCheckName:ReadinessCheckOutput', readinessCheckOutput_readinessCheckName - Name for a ReadinessCheck

$sel:tags:ReadinessCheckOutput', readinessCheckOutput_tags - Undocumented member.

$sel:readinessCheckArn:ReadinessCheckOutput', readinessCheckOutput_readinessCheckArn - Arn associated with ReadinessCheck

$sel:resourceSet:ReadinessCheckOutput', readinessCheckOutput_resourceSet - Name of the ResourceSet to be checked

readinessCheckOutput_readinessCheckName :: Lens' ReadinessCheckOutput (Maybe Text)

Name for a ReadinessCheck

readinessCheckOutput_tags :: Lens' ReadinessCheckOutput (Maybe (HashMap Text Text))

Undocumented member.

readinessCheckOutput_readinessCheckArn :: Lens' ReadinessCheckOutput Text

Arn associated with ReadinessCheck

readinessCheckOutput_resourceSet :: Lens' ReadinessCheckOutput Text

Name of the ResourceSet to be checked

ReadinessCheckSummary

data ReadinessCheckSummary

Summary of ReadinessCheck status, paginated in GetRecoveryGroupReadinessSummary and GetCellReadinessSummary

See: newReadinessCheckSummary smart constructor.

Constructors

ReadinessCheckSummary'

Fields

readiness :: Maybe Readiness

The readiness of this ReadinessCheck

readinessCheckName :: Maybe Text

The name of a ReadinessCheck which is part of the given RecoveryGroup or Cell

newReadinessCheckSummary :: ReadinessCheckSummary

Create a value of ReadinessCheckSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:readiness:ReadinessCheckSummary', readinessCheckSummary_readiness - The readiness of this ReadinessCheck

$sel:readinessCheckName:ReadinessCheckSummary', readinessCheckSummary_readinessCheckName - The name of a ReadinessCheck which is part of the given RecoveryGroup or Cell

readinessCheckSummary_readiness :: Lens' ReadinessCheckSummary (Maybe Readiness)

The readiness of this ReadinessCheck

readinessCheckSummary_readinessCheckName :: Lens' ReadinessCheckSummary (Maybe Text)

The name of a ReadinessCheck which is part of the given RecoveryGroup or Cell

Recommendation

data Recommendation

Guidance for improving Recovery Group resilliancy

See: newRecommendation smart constructor.

Constructors

Recommendation'

Fields

recommendationText :: Text

Guidance text for recommendation

newRecommendation

Arguments

:: Text
$sel:recommendationText:Recommendation'

-> Recommendation

Create a value of Recommendation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recommendationText:Recommendation', recommendation_recommendationText - Guidance text for recommendation

recommendation_recommendationText :: Lens' Recommendation Text

Guidance text for recommendation

RecoveryGroupOutput

data RecoveryGroupOutput

A Recovery Group generally containing multiple Cells

See: newRecoveryGroupOutput smart constructor.

Constructors

RecoveryGroupOutput'

Fields

tags :: Maybe (HashMap Text Text)

recoveryGroupArn :: Text

The arn for the RecoveryGroup

recoveryGroupName :: Text

The name of the RecoveryGroup

cells :: [Text]

A list of Cell arns

newRecoveryGroupOutput

Arguments

:: Text
$sel:recoveryGroupArn:RecoveryGroupOutput'

-> Text
$sel:recoveryGroupName:RecoveryGroupOutput'

-> RecoveryGroupOutput

Create a value of RecoveryGroupOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:RecoveryGroupOutput', recoveryGroupOutput_tags - Undocumented member.

$sel:recoveryGroupArn:RecoveryGroupOutput', recoveryGroupOutput_recoveryGroupArn - The arn for the RecoveryGroup

$sel:recoveryGroupName:RecoveryGroupOutput', recoveryGroupOutput_recoveryGroupName - The name of the RecoveryGroup

$sel:cells:RecoveryGroupOutput', recoveryGroupOutput_cells - A list of Cell arns

recoveryGroupOutput_tags :: Lens' RecoveryGroupOutput (Maybe (HashMap Text Text))

Undocumented member.

recoveryGroupOutput_recoveryGroupArn :: Lens' RecoveryGroupOutput Text

The arn for the RecoveryGroup

recoveryGroupOutput_recoveryGroupName :: Lens' RecoveryGroupOutput Text

The name of the RecoveryGroup

recoveryGroupOutput_cells :: Lens' RecoveryGroupOutput [Text]

A list of Cell arns

Resource

data Resource

The resource element of a ResourceSet

See: newResource smart constructor.

Constructors

Resource'

Fields

readinessScopes :: Maybe [Text]

A list of RecoveryGroup ARNs and/or Cell ARNs that this resource is contained within.

resourceArn :: Maybe Text

The ARN of the AWS resource, can be skipped if dnsTargetResource is used

componentId :: Maybe Text

The component id of the resource, generated by the service when dnsTargetResource is used

dnsTargetResource :: Maybe DNSTargetResource

newResource :: Resource

Create a value of Resource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:readinessScopes:Resource', resource_readinessScopes - A list of RecoveryGroup ARNs and/or Cell ARNs that this resource is contained within.

$sel:resourceArn:Resource', resource_resourceArn - The ARN of the AWS resource, can be skipped if dnsTargetResource is used

$sel:componentId:Resource', resource_componentId - The component id of the resource, generated by the service when dnsTargetResource is used

$sel:dnsTargetResource:Resource', resource_dnsTargetResource - Undocumented member.

resource_readinessScopes :: Lens' Resource (Maybe [Text])

A list of RecoveryGroup ARNs and/or Cell ARNs that this resource is contained within.

resource_resourceArn :: Lens' Resource (Maybe Text)

The ARN of the AWS resource, can be skipped if dnsTargetResource is used

resource_componentId :: Lens' Resource (Maybe Text)

The component id of the resource, generated by the service when dnsTargetResource is used

resource_dnsTargetResource :: Lens' Resource (Maybe DNSTargetResource)

Undocumented member.

ResourceResult

data ResourceResult

Result with status for an individual resource.

See: newResourceResult smart constructor.

Constructors

ResourceResult'

Fields

resourceArn :: Maybe Text

The ARN of the resource

componentId :: Maybe Text

The component id of the resource

readiness :: Readiness

The readiness of the resource.

lastCheckedTimestamp :: POSIX

The time the resource was last checked for readiness, in ISO-8601 format, UTC.

newResourceResult

Arguments

:: Readiness
$sel:readiness:ResourceResult'

-> UTCTime
$sel:lastCheckedTimestamp:ResourceResult'

-> ResourceResult

Create a value of ResourceResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ResourceResult', resourceResult_resourceArn - The ARN of the resource

$sel:componentId:ResourceResult', resourceResult_componentId - The component id of the resource

$sel:readiness:ResourceResult', resourceResult_readiness - The readiness of the resource.

$sel:lastCheckedTimestamp:ResourceResult', resourceResult_lastCheckedTimestamp - The time the resource was last checked for readiness, in ISO-8601 format, UTC.

resourceResult_resourceArn :: Lens' ResourceResult (Maybe Text)

The ARN of the resource

resourceResult_componentId :: Lens' ResourceResult (Maybe Text)

The component id of the resource

resourceResult_readiness :: Lens' ResourceResult Readiness

The readiness of the resource.

resourceResult_lastCheckedTimestamp :: Lens' ResourceResult UTCTime

The time the resource was last checked for readiness, in ISO-8601 format, UTC.

ResourceSetOutput

data ResourceSetOutput

A collection of resources of the same type

See: newResourceSetOutput smart constructor.

Constructors

ResourceSetOutput'

Fields

tags :: Maybe (HashMap Text Text)

resourceSetType :: Text

AWS Resource Type of the resources in the ResourceSet

resourceSetName :: Text

The name of the ResourceSet

resourceSetArn :: Text

The arn for the ResourceSet

resources :: [Resource]

A list of Resource objects

newResourceSetOutput

Arguments

:: Text
$sel:resourceSetType:ResourceSetOutput'

-> Text
$sel:resourceSetName:ResourceSetOutput'

-> Text
$sel:resourceSetArn:ResourceSetOutput'

-> ResourceSetOutput

Create a value of ResourceSetOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ResourceSetOutput', resourceSetOutput_tags - Undocumented member.

$sel:resourceSetType:ResourceSetOutput', resourceSetOutput_resourceSetType - AWS Resource Type of the resources in the ResourceSet

$sel:resourceSetName:ResourceSetOutput', resourceSetOutput_resourceSetName - The name of the ResourceSet

$sel:resourceSetArn:ResourceSetOutput', resourceSetOutput_resourceSetArn - The arn for the ResourceSet

$sel:resources:ResourceSetOutput', resourceSetOutput_resources - A list of Resource objects

resourceSetOutput_tags :: Lens' ResourceSetOutput (Maybe (HashMap Text Text))

Undocumented member.

resourceSetOutput_resourceSetType :: Lens' ResourceSetOutput Text

AWS Resource Type of the resources in the ResourceSet

resourceSetOutput_resourceSetName :: Lens' ResourceSetOutput Text

The name of the ResourceSet

resourceSetOutput_resourceSetArn :: Lens' ResourceSetOutput Text

The arn for the ResourceSet

resourceSetOutput_resources :: Lens' ResourceSetOutput [Resource]

A list of Resource objects

RuleResult

data RuleResult

Result with status for an individual rule..

See: newRuleResult smart constructor.

Constructors

RuleResult'

Fields

messages :: [Message]

Details about the resource's readiness

readiness :: Readiness

The readiness at rule level.

ruleId :: Text

The identifier of the rule.

lastCheckedTimestamp :: POSIX

The time the resource was last checked for readiness, in ISO-8601 format, UTC.

newRuleResult

Arguments

:: Readiness
$sel:readiness:RuleResult'

-> Text
$sel:ruleId:RuleResult'

-> UTCTime
$sel:lastCheckedTimestamp:RuleResult'

-> RuleResult

Create a value of RuleResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:messages:RuleResult', ruleResult_messages - Details about the resource's readiness

$sel:readiness:RuleResult', ruleResult_readiness - The readiness at rule level.

$sel:ruleId:RuleResult', ruleResult_ruleId - The identifier of the rule.

$sel:lastCheckedTimestamp:RuleResult', ruleResult_lastCheckedTimestamp - The time the resource was last checked for readiness, in ISO-8601 format, UTC.

ruleResult_messages :: Lens' RuleResult [Message]

Details about the resource's readiness

ruleResult_readiness :: Lens' RuleResult Readiness

The readiness at rule level.

ruleResult_ruleId :: Lens' RuleResult Text

The identifier of the rule.

ruleResult_lastCheckedTimestamp :: Lens' RuleResult UTCTime

The time the resource was last checked for readiness, in ISO-8601 format, UTC.

TargetResource

data TargetResource

The target resource the R53 record points to

See: newTargetResource smart constructor.

Constructors

TargetResource'

Fields

r53Resource :: Maybe R53ResourceRecord

nLBResource :: Maybe NLBResource

newTargetResource :: TargetResource

Create a value of TargetResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:r53Resource:TargetResource', targetResource_r53Resource - Undocumented member.

$sel:nLBResource:TargetResource', targetResource_nLBResource - Undocumented member.

targetResource_r53Resource :: Lens' TargetResource (Maybe R53ResourceRecord)

Undocumented member.

targetResource_nLBResource :: Lens' TargetResource (Maybe NLBResource)

Undocumented member.