Network.AWS.Route53RecoveryReadiness.Types.ReadinessCheckOutput

Description

Documentation

data ReadinessCheckOutput

A resource used for checking the readiness of a Resource Set

See: newReadinessCheckOutput smart constructor.

Constructors

ReadinessCheckOutput'

Fields

readinessCheckName :: Maybe Text

Name for a ReadinessCheck

tags :: Maybe (HashMap Text Text)

readinessCheckArn :: Text

Arn associated with ReadinessCheck

resourceSet :: Text

Name of the ResourceSet to be checked

newReadinessCheckOutput

Arguments

:: Text
$sel:readinessCheckArn:ReadinessCheckOutput'

-> Text
$sel:resourceSet:ReadinessCheckOutput'

-> ReadinessCheckOutput

Create a value of ReadinessCheckOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:readinessCheckName:ReadinessCheckOutput', readinessCheckOutput_readinessCheckName - Name for a ReadinessCheck

$sel:tags:ReadinessCheckOutput', readinessCheckOutput_tags - Undocumented member.

$sel:readinessCheckArn:ReadinessCheckOutput', readinessCheckOutput_readinessCheckArn - Arn associated with ReadinessCheck

$sel:resourceSet:ReadinessCheckOutput', readinessCheckOutput_resourceSet - Name of the ResourceSet to be checked

readinessCheckOutput_readinessCheckName :: Lens' ReadinessCheckOutput (Maybe Text)

Name for a ReadinessCheck

readinessCheckOutput_tags :: Lens' ReadinessCheckOutput (Maybe (HashMap Text Text))

Undocumented member.

readinessCheckOutput_readinessCheckArn :: Lens' ReadinessCheckOutput Text

Arn associated with ReadinessCheck

readinessCheckOutput_resourceSet :: Lens' ReadinessCheckOutput Text

Name of the ResourceSet to be checked