Network.AWS.Route53RecoveryReadiness.UpdateCell

Description

Updates an existing Cell.

Creating a Request

data UpdateCell

Parameters to update for the Cell

See: newUpdateCell smart constructor.

Constructors

UpdateCell'

Fields

cellName :: Text

The Cell to update

cells :: [Text]

A list of Cell arns, completely replaces previous list

newUpdateCell

Arguments

:: Text
$sel:cellName:UpdateCell'

-> UpdateCell

Create a value of UpdateCell with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cellName:UpdateCell', updateCell_cellName - The Cell to update

$sel:cells:UpdateCell', updateCell_cells - A list of Cell arns, completely replaces previous list

Request Lenses

updateCell_cellName :: Lens' UpdateCell Text

The Cell to update

updateCell_cells :: Lens' UpdateCell [Text]

A list of Cell arns, completely replaces previous list

Destructuring the Response

data UpdateCellResponse

See: newUpdateCellResponse smart constructor.

Constructors

UpdateCellResponse'

Fields

cells :: Maybe [Text]

A list of Cell arns

parentReadinessScopes :: Maybe [Text]

A list of Cell ARNs and/or RecoveryGroup ARNs

cellName :: Maybe Text

The name of the Cell

cellArn :: Maybe Text

The arn for the Cell

tags :: Maybe (HashMap Text Text)

httpStatus :: Int

The response's http status code.

newUpdateCellResponse

Arguments

:: Int
$sel:httpStatus:UpdateCellResponse'

-> UpdateCellResponse

Create a value of UpdateCellResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cells:UpdateCell', updateCellResponse_cells - A list of Cell arns

$sel:parentReadinessScopes:UpdateCellResponse', updateCellResponse_parentReadinessScopes - A list of Cell ARNs and/or RecoveryGroup ARNs

$sel:cellName:UpdateCell', updateCellResponse_cellName - The name of the Cell

$sel:cellArn:UpdateCellResponse', updateCellResponse_cellArn - The arn for the Cell

$sel:tags:UpdateCellResponse', updateCellResponse_tags - Undocumented member.

$sel:httpStatus:UpdateCellResponse', updateCellResponse_httpStatus - The response's http status code.

Response Lenses

updateCellResponse_cells :: Lens' UpdateCellResponse (Maybe [Text])

A list of Cell arns

updateCellResponse_parentReadinessScopes :: Lens' UpdateCellResponse (Maybe [Text])

A list of Cell ARNs and/or RecoveryGroup ARNs

updateCellResponse_cellName :: Lens' UpdateCellResponse (Maybe Text)

The name of the Cell

updateCellResponse_cellArn :: Lens' UpdateCellResponse (Maybe Text)

The arn for the Cell

updateCellResponse_tags :: Lens' UpdateCellResponse (Maybe (HashMap Text Text))

Undocumented member.

updateCellResponse_httpStatus :: Lens' UpdateCellResponse Int

The response's http status code.