Network.AWS.Route53RecoveryReadiness.UpdateRecoveryGroup

Description

Updates an existing Recovery Group.

Creating a Request

data UpdateRecoveryGroup

Parameters to update for the RecoveryGroup

See: newUpdateRecoveryGroup smart constructor.

Constructors

UpdateRecoveryGroup'

Fields

recoveryGroupName :: Text

The RecoveryGroup to update

cells :: [Text]

A list of Cell arns, completely replaces previous list

newUpdateRecoveryGroup

Arguments

:: Text
$sel:recoveryGroupName:UpdateRecoveryGroup'

-> UpdateRecoveryGroup

Create a value of UpdateRecoveryGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recoveryGroupName:UpdateRecoveryGroup', updateRecoveryGroup_recoveryGroupName - The RecoveryGroup to update

$sel:cells:UpdateRecoveryGroup', updateRecoveryGroup_cells - A list of Cell arns, completely replaces previous list

Request Lenses

updateRecoveryGroup_recoveryGroupName :: Lens' UpdateRecoveryGroup Text

The RecoveryGroup to update

updateRecoveryGroup_cells :: Lens' UpdateRecoveryGroup [Text]

A list of Cell arns, completely replaces previous list

Destructuring the Response

data UpdateRecoveryGroupResponse

See: newUpdateRecoveryGroupResponse smart constructor.

Constructors

UpdateRecoveryGroupResponse'

Fields

cells :: Maybe [Text]

A list of Cell arns

recoveryGroupName :: Maybe Text

The name of the RecoveryGroup

recoveryGroupArn :: Maybe Text

The arn for the RecoveryGroup

tags :: Maybe (HashMap Text Text)

httpStatus :: Int

The response's http status code.

newUpdateRecoveryGroupResponse

Arguments

:: Int
$sel:httpStatus:UpdateRecoveryGroupResponse'

-> UpdateRecoveryGroupResponse

Create a value of UpdateRecoveryGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cells:UpdateRecoveryGroup', updateRecoveryGroupResponse_cells - A list of Cell arns

$sel:recoveryGroupName:UpdateRecoveryGroup', updateRecoveryGroupResponse_recoveryGroupName - The name of the RecoveryGroup

$sel:recoveryGroupArn:UpdateRecoveryGroupResponse', updateRecoveryGroupResponse_recoveryGroupArn - The arn for the RecoveryGroup

$sel:tags:UpdateRecoveryGroupResponse', updateRecoveryGroupResponse_tags - Undocumented member.

$sel:httpStatus:UpdateRecoveryGroupResponse', updateRecoveryGroupResponse_httpStatus - The response's http status code.

Response Lenses

updateRecoveryGroupResponse_cells :: Lens' UpdateRecoveryGroupResponse (Maybe [Text])

A list of Cell arns

updateRecoveryGroupResponse_recoveryGroupName :: Lens' UpdateRecoveryGroupResponse (Maybe Text)

The name of the RecoveryGroup

updateRecoveryGroupResponse_recoveryGroupArn :: Lens' UpdateRecoveryGroupResponse (Maybe Text)

The arn for the RecoveryGroup

updateRecoveryGroupResponse_tags :: Lens' UpdateRecoveryGroupResponse (Maybe (HashMap Text Text))

Undocumented member.

updateRecoveryGroupResponse_httpStatus :: Lens' UpdateRecoveryGroupResponse Int

The response's http status code.