Network.AWS.Route53Resolver.AssociateResolverRule

Description

Associates a Resolver rule with a VPC. When you associate a rule with a VPC, Resolver forwards all DNS queries for the domain name that is specified in the rule and that originate in the VPC. The queries are forwarded to the IP addresses for the DNS resolvers that are specified in the rule. For more information about rules, see CreateResolverRule.

Creating a Request

data AssociateResolverRule

See: newAssociateResolverRule smart constructor.

Constructors

AssociateResolverRule'

Fields

name :: Maybe Text

A name for the association that you're creating between a Resolver rule and a VPC.

resolverRuleId :: Text

The ID of the Resolver rule that you want to associate with the VPC. To list the existing Resolver rules, use ListResolverRules.

vPCId :: Text

The ID of the VPC that you want to associate the Resolver rule with.

newAssociateResolverRule

Arguments

:: Text
$sel:resolverRuleId:AssociateResolverRule'

-> Text
$sel:vPCId:AssociateResolverRule'

-> AssociateResolverRule

Create a value of AssociateResolverRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:AssociateResolverRule', associateResolverRule_name - A name for the association that you're creating between a Resolver rule and a VPC.

$sel:resolverRuleId:AssociateResolverRule', associateResolverRule_resolverRuleId - The ID of the Resolver rule that you want to associate with the VPC. To list the existing Resolver rules, use ListResolverRules.

$sel:vPCId:AssociateResolverRule', associateResolverRule_vPCId - The ID of the VPC that you want to associate the Resolver rule with.

Request Lenses

associateResolverRule_name :: Lens' AssociateResolverRule (Maybe Text)

A name for the association that you're creating between a Resolver rule and a VPC.

associateResolverRule_resolverRuleId :: Lens' AssociateResolverRule Text

The ID of the Resolver rule that you want to associate with the VPC. To list the existing Resolver rules, use ListResolverRules.

associateResolverRule_vPCId :: Lens' AssociateResolverRule Text

The ID of the VPC that you want to associate the Resolver rule with.

Destructuring the Response

data AssociateResolverRuleResponse

See: newAssociateResolverRuleResponse smart constructor.

Constructors

AssociateResolverRuleResponse'

Fields

resolverRuleAssociation :: Maybe ResolverRuleAssociation

Information about the AssociateResolverRule request, including the status of the request.

httpStatus :: Int

The response's http status code.

newAssociateResolverRuleResponse

Arguments

:: Int
$sel:httpStatus:AssociateResolverRuleResponse'

-> AssociateResolverRuleResponse

Create a value of AssociateResolverRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolverRuleAssociation:AssociateResolverRuleResponse', associateResolverRuleResponse_resolverRuleAssociation - Information about the AssociateResolverRule request, including the status of the request.

$sel:httpStatus:AssociateResolverRuleResponse', associateResolverRuleResponse_httpStatus - The response's http status code.

Response Lenses

associateResolverRuleResponse_resolverRuleAssociation :: Lens' AssociateResolverRuleResponse (Maybe ResolverRuleAssociation)

Information about the AssociateResolverRule request, including the status of the request.

associateResolverRuleResponse_httpStatus :: Lens' AssociateResolverRuleResponse Int

The response's http status code.