Network.AWS.Route53Resolver.CreateResolverEndpoint

Description

Creates a Resolver endpoint. There are two types of Resolver endpoints, inbound and outbound:

An inbound Resolver endpoint forwards DNS queries to the DNS service for a VPC from your network.

An outbound Resolver endpoint forwards DNS queries from the DNS service for a VPC to your network.

Creating a Request

data CreateResolverEndpoint

See: newCreateResolverEndpoint smart constructor.

Constructors

CreateResolverEndpoint'

Fields

name :: Maybe Text

A friendly name that lets you easily find a configuration in the Resolver dashboard in the Route 53 console.

tags :: Maybe [Tag]

A list of the tag keys and values that you want to associate with the endpoint.

creatorRequestId :: Text

A unique string that identifies the request and that allows failed requests to be retried without the risk of running the operation twice. CreatorRequestId can be any unique string, for example, a date/time stamp.

securityGroupIds :: [Text]

The ID of one or more security groups that you want to use to control access to this VPC. The security group that you specify must include one or more inbound rules (for inbound Resolver endpoints) or outbound rules (for outbound Resolver endpoints). Inbound and outbound rules must allow TCP and UDP access. For inbound access, open port 53. For outbound access, open the port that you're using for DNS queries on your network.

direction :: ResolverEndpointDirection

Specify the applicable value:

INBOUND: Resolver forwards DNS queries to the DNS service for a VPC from your network

OUTBOUND: Resolver forwards DNS queries from the DNS service for a VPC to your network

ipAddresses :: NonEmpty IpAddressRequest

The subnets and IP addresses in your VPC that DNS queries originate from (for outbound endpoints) or that you forward DNS queries to (for inbound endpoints). The subnet ID uniquely identifies a VPC.

newCreateResolverEndpoint

Arguments

:: Text
$sel:creatorRequestId:CreateResolverEndpoint'

-> ResolverEndpointDirection
$sel:direction:CreateResolverEndpoint'

-> NonEmpty IpAddressRequest
$sel:ipAddresses:CreateResolverEndpoint'

-> CreateResolverEndpoint

Create a value of CreateResolverEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreateResolverEndpoint', createResolverEndpoint_name - A friendly name that lets you easily find a configuration in the Resolver dashboard in the Route 53 console.

$sel:tags:CreateResolverEndpoint', createResolverEndpoint_tags - A list of the tag keys and values that you want to associate with the endpoint.

$sel:creatorRequestId:CreateResolverEndpoint', createResolverEndpoint_creatorRequestId - A unique string that identifies the request and that allows failed requests to be retried without the risk of running the operation twice. CreatorRequestId can be any unique string, for example, a date/time stamp.

$sel:securityGroupIds:CreateResolverEndpoint', createResolverEndpoint_securityGroupIds - The ID of one or more security groups that you want to use to control access to this VPC. The security group that you specify must include one or more inbound rules (for inbound Resolver endpoints) or outbound rules (for outbound Resolver endpoints). Inbound and outbound rules must allow TCP and UDP access. For inbound access, open port 53. For outbound access, open the port that you're using for DNS queries on your network.

$sel:direction:CreateResolverEndpoint', createResolverEndpoint_direction - Specify the applicable value:

INBOUND: Resolver forwards DNS queries to the DNS service for a VPC from your network

OUTBOUND: Resolver forwards DNS queries from the DNS service for a VPC to your network

$sel:ipAddresses:CreateResolverEndpoint', createResolverEndpoint_ipAddresses - The subnets and IP addresses in your VPC that DNS queries originate from (for outbound endpoints) or that you forward DNS queries to (for inbound endpoints). The subnet ID uniquely identifies a VPC.

Request Lenses

createResolverEndpoint_name :: Lens' CreateResolverEndpoint (Maybe Text)

A friendly name that lets you easily find a configuration in the Resolver dashboard in the Route 53 console.

createResolverEndpoint_tags :: Lens' CreateResolverEndpoint (Maybe [Tag])

A list of the tag keys and values that you want to associate with the endpoint.

createResolverEndpoint_creatorRequestId :: Lens' CreateResolverEndpoint Text

A unique string that identifies the request and that allows failed requests to be retried without the risk of running the operation twice. CreatorRequestId can be any unique string, for example, a date/time stamp.

createResolverEndpoint_securityGroupIds :: Lens' CreateResolverEndpoint [Text]

The ID of one or more security groups that you want to use to control access to this VPC. The security group that you specify must include one or more inbound rules (for inbound Resolver endpoints) or outbound rules (for outbound Resolver endpoints). Inbound and outbound rules must allow TCP and UDP access. For inbound access, open port 53. For outbound access, open the port that you're using for DNS queries on your network.

createResolverEndpoint_direction :: Lens' CreateResolverEndpoint ResolverEndpointDirection

Specify the applicable value:

INBOUND: Resolver forwards DNS queries to the DNS service for a VPC from your network

OUTBOUND: Resolver forwards DNS queries from the DNS service for a VPC to your network

createResolverEndpoint_ipAddresses :: Lens' CreateResolverEndpoint (NonEmpty IpAddressRequest)

The subnets and IP addresses in your VPC that DNS queries originate from (for outbound endpoints) or that you forward DNS queries to (for inbound endpoints). The subnet ID uniquely identifies a VPC.

Destructuring the Response

data CreateResolverEndpointResponse

See: newCreateResolverEndpointResponse smart constructor.

Constructors

CreateResolverEndpointResponse'

Fields

resolverEndpoint :: Maybe ResolverEndpoint

Information about the CreateResolverEndpoint request, including the status of the request.

httpStatus :: Int

The response's http status code.

newCreateResolverEndpointResponse

Arguments

:: Int
$sel:httpStatus:CreateResolverEndpointResponse'

-> CreateResolverEndpointResponse

Create a value of CreateResolverEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolverEndpoint:CreateResolverEndpointResponse', createResolverEndpointResponse_resolverEndpoint - Information about the CreateResolverEndpoint request, including the status of the request.

$sel:httpStatus:CreateResolverEndpointResponse', createResolverEndpointResponse_httpStatus - The response's http status code.

Response Lenses

createResolverEndpointResponse_resolverEndpoint :: Lens' CreateResolverEndpointResponse (Maybe ResolverEndpoint)

Information about the CreateResolverEndpoint request, including the status of the request.

createResolverEndpointResponse_httpStatus :: Lens' CreateResolverEndpointResponse Int

The response's http status code.