Network.AWS.Route53Resolver.GetFirewallDomainList

Description

Retrieves the specified firewall domain list.

Creating a Request

data GetFirewallDomainList

See: newGetFirewallDomainList smart constructor.

Constructors

GetFirewallDomainList'

Fields

firewallDomainListId :: Text

The ID of the domain list.

newGetFirewallDomainList

Arguments

:: Text
$sel:firewallDomainListId:GetFirewallDomainList'

-> GetFirewallDomainList

Create a value of GetFirewallDomainList with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:firewallDomainListId:GetFirewallDomainList', getFirewallDomainList_firewallDomainListId - The ID of the domain list.

Request Lenses

getFirewallDomainList_firewallDomainListId :: Lens' GetFirewallDomainList Text

The ID of the domain list.

Destructuring the Response

data GetFirewallDomainListResponse

See: newGetFirewallDomainListResponse smart constructor.

Constructors

GetFirewallDomainListResponse'

Fields

firewallDomainList :: Maybe FirewallDomainList

The domain list that you requested.

httpStatus :: Int

The response's http status code.

newGetFirewallDomainListResponse

Arguments

:: Int
$sel:httpStatus:GetFirewallDomainListResponse'

-> GetFirewallDomainListResponse

Create a value of GetFirewallDomainListResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:firewallDomainList:GetFirewallDomainListResponse', getFirewallDomainListResponse_firewallDomainList - The domain list that you requested.

$sel:httpStatus:GetFirewallDomainListResponse', getFirewallDomainListResponse_httpStatus - The response's http status code.

Response Lenses

getFirewallDomainListResponse_firewallDomainList :: Lens' GetFirewallDomainListResponse (Maybe FirewallDomainList)

The domain list that you requested.

getFirewallDomainListResponse_httpStatus :: Lens' GetFirewallDomainListResponse Int

The response's http status code.