Network.AWS.Route53Resolver.GetFirewallRuleGroupAssociation

Description

Retrieves a firewall rule group association, which enables DNS filtering for a VPC with one rule group. A VPC can have more than one firewall rule group association, and a rule group can be associated with more than one VPC.

Creating a Request

data GetFirewallRuleGroupAssociation

See: newGetFirewallRuleGroupAssociation smart constructor.

Constructors

GetFirewallRuleGroupAssociation'

Fields

firewallRuleGroupAssociationId :: Text

The identifier of the FirewallRuleGroupAssociation.

newGetFirewallRuleGroupAssociation

Arguments

:: Text
$sel:firewallRuleGroupAssociationId:GetFirewallRuleGroupAssociation'

-> GetFirewallRuleGroupAssociation

Create a value of GetFirewallRuleGroupAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:firewallRuleGroupAssociationId:GetFirewallRuleGroupAssociation', getFirewallRuleGroupAssociation_firewallRuleGroupAssociationId - The identifier of the FirewallRuleGroupAssociation.

Request Lenses

getFirewallRuleGroupAssociation_firewallRuleGroupAssociationId :: Lens' GetFirewallRuleGroupAssociation Text

The identifier of the FirewallRuleGroupAssociation.

Destructuring the Response

data GetFirewallRuleGroupAssociationResponse

See: newGetFirewallRuleGroupAssociationResponse smart constructor.

Constructors

GetFirewallRuleGroupAssociationResponse'

Fields

firewallRuleGroupAssociation :: Maybe FirewallRuleGroupAssociation

The association that you requested.

httpStatus :: Int

The response's http status code.

newGetFirewallRuleGroupAssociationResponse

Arguments

:: Int
$sel:httpStatus:GetFirewallRuleGroupAssociationResponse'

-> GetFirewallRuleGroupAssociationResponse

Create a value of GetFirewallRuleGroupAssociationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:firewallRuleGroupAssociation:GetFirewallRuleGroupAssociationResponse', getFirewallRuleGroupAssociationResponse_firewallRuleGroupAssociation - The association that you requested.

$sel:httpStatus:GetFirewallRuleGroupAssociationResponse', getFirewallRuleGroupAssociationResponse_httpStatus - The response's http status code.

Response Lenses

getFirewallRuleGroupAssociationResponse_firewallRuleGroupAssociation :: Lens' GetFirewallRuleGroupAssociationResponse (Maybe FirewallRuleGroupAssociation)

The association that you requested.

getFirewallRuleGroupAssociationResponse_httpStatus :: Lens' GetFirewallRuleGroupAssociationResponse Int

The response's http status code.