Network.AWS.Route53Resolver.GetResolverDnssecConfig

Description

Gets DNSSEC validation information for a specified resource.

Creating a Request

data GetResolverDnssecConfig

See: newGetResolverDnssecConfig smart constructor.

Constructors

GetResolverDnssecConfig'

Fields

resourceId :: Text

The ID of the virtual private cloud (VPC) for the DNSSEC validation status.

newGetResolverDnssecConfig

Arguments

:: Text
$sel:resourceId:GetResolverDnssecConfig'

-> GetResolverDnssecConfig

Create a value of GetResolverDnssecConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:GetResolverDnssecConfig', getResolverDnssecConfig_resourceId - The ID of the virtual private cloud (VPC) for the DNSSEC validation status.

Request Lenses

getResolverDnssecConfig_resourceId :: Lens' GetResolverDnssecConfig Text

The ID of the virtual private cloud (VPC) for the DNSSEC validation status.

Destructuring the Response

data GetResolverDnssecConfigResponse

See: newGetResolverDnssecConfigResponse smart constructor.

Constructors

GetResolverDnssecConfigResponse'

Fields

resolverDNSSECConfig :: Maybe ResolverDnssecConfig

The information about a configuration for DNSSEC validation.

httpStatus :: Int

The response's http status code.

newGetResolverDnssecConfigResponse

Arguments

:: Int
$sel:httpStatus:GetResolverDnssecConfigResponse'

-> GetResolverDnssecConfigResponse

Create a value of GetResolverDnssecConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolverDNSSECConfig:GetResolverDnssecConfigResponse', getResolverDnssecConfigResponse_resolverDNSSECConfig - The information about a configuration for DNSSEC validation.

$sel:httpStatus:GetResolverDnssecConfigResponse', getResolverDnssecConfigResponse_httpStatus - The response's http status code.

Response Lenses

getResolverDnssecConfigResponse_resolverDNSSECConfig :: Lens' GetResolverDnssecConfigResponse (Maybe ResolverDnssecConfig)

The information about a configuration for DNSSEC validation.

getResolverDnssecConfigResponse_httpStatus :: Lens' GetResolverDnssecConfigResponse Int

The response's http status code.