Network.AWS.Route53Resolver.GetResolverQueryLogConfigAssociation

Description

Gets information about a specified association between a Resolver query logging configuration and an Amazon VPC. When you associate a VPC with a query logging configuration, Resolver logs DNS queries that originate in that VPC.

Creating a Request

data GetResolverQueryLogConfigAssociation

See: newGetResolverQueryLogConfigAssociation smart constructor.

Constructors

GetResolverQueryLogConfigAssociation'

Fields

resolverQueryLogConfigAssociationId :: Text

The ID of the Resolver query logging configuration association that you want to get information about.

newGetResolverQueryLogConfigAssociation

Arguments

:: Text
$sel:resolverQueryLogConfigAssociationId:GetResolverQueryLogConfigAssociation'

-> GetResolverQueryLogConfigAssociation

Create a value of GetResolverQueryLogConfigAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolverQueryLogConfigAssociationId:GetResolverQueryLogConfigAssociation', getResolverQueryLogConfigAssociation_resolverQueryLogConfigAssociationId - The ID of the Resolver query logging configuration association that you want to get information about.

Request Lenses

getResolverQueryLogConfigAssociation_resolverQueryLogConfigAssociationId :: Lens' GetResolverQueryLogConfigAssociation Text

The ID of the Resolver query logging configuration association that you want to get information about.

Destructuring the Response

data GetResolverQueryLogConfigAssociationResponse

See: newGetResolverQueryLogConfigAssociationResponse smart constructor.

Constructors

GetResolverQueryLogConfigAssociationResponse'

Fields

resolverQueryLogConfigAssociation :: Maybe ResolverQueryLogConfigAssociation

Information about the Resolver query logging configuration association that you specified in a GetQueryLogConfigAssociation request.

httpStatus :: Int

The response's http status code.

newGetResolverQueryLogConfigAssociationResponse

Arguments

:: Int
$sel:httpStatus:GetResolverQueryLogConfigAssociationResponse'

-> GetResolverQueryLogConfigAssociationResponse

Create a value of GetResolverQueryLogConfigAssociationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolverQueryLogConfigAssociation:GetResolverQueryLogConfigAssociationResponse', getResolverQueryLogConfigAssociationResponse_resolverQueryLogConfigAssociation - Information about the Resolver query logging configuration association that you specified in a GetQueryLogConfigAssociation request.

$sel:httpStatus:GetResolverQueryLogConfigAssociationResponse', getResolverQueryLogConfigAssociationResponse_httpStatus - The response's http status code.

Response Lenses

getResolverQueryLogConfigAssociationResponse_resolverQueryLogConfigAssociation :: Lens' GetResolverQueryLogConfigAssociationResponse (Maybe ResolverQueryLogConfigAssociation)

Information about the Resolver query logging configuration association that you specified in a GetQueryLogConfigAssociation request.

getResolverQueryLogConfigAssociationResponse_httpStatus :: Lens' GetResolverQueryLogConfigAssociationResponse Int

The response's http status code.