Network.AWS.Route53Resolver.ListFirewallRules

Description

Retrieves the firewall rules that you have defined for the specified firewall rule group. DNS Firewall uses the rules in a rule group to filter DNS network traffic for a VPC.

A single call might return only a partial list of the rules. For information, see MaxResults.

This operation returns paginated results.

Creating a Request

data ListFirewallRules

See: newListFirewallRules smart constructor.

Constructors

ListFirewallRules'

Fields

priority :: Maybe Int

Optional additional filter for the rules to retrieve.

The setting that determines the processing order of the rules in a rule group. DNS Firewall processes the rules in a rule group by order of priority, starting from the lowest setting.

action :: Maybe Action

Optional additional filter for the rules to retrieve.

The action that DNS Firewall should take on a DNS query when it matches one of the domains in the rule's domain list:

ALLOW - Permit the request to go through.

ALERT - Permit the request to go through but send an alert to the logs.

BLOCK - Disallow the request. If this is specified, additional handling details are provided in the rule's BlockResponse setting.

nextToken :: Maybe Text

For the first call to this list request, omit this value.

When you request a list of objects, Resolver returns at most the number of objects specified in MaxResults. If more objects are available for retrieval, Resolver returns a NextToken value in the response. To retrieve the next batch of objects, use the token that was returned for the prior request in your next request.

maxResults :: Maybe Natural

The maximum number of objects that you want Resolver to return for this request. If more objects are available, in the response, Resolver provides a NextToken value that you can use in a subsequent call to get the next batch of objects.

If you don't specify a value for MaxResults, Resolver returns up to 100 objects.

firewallRuleGroupId :: Text

The unique identifier of the firewall rule group that you want to retrieve the rules for.

newListFirewallRules

Arguments

:: Text
$sel:firewallRuleGroupId:ListFirewallRules'

-> ListFirewallRules

Create a value of ListFirewallRules with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:priority:ListFirewallRules', listFirewallRules_priority - Optional additional filter for the rules to retrieve.

The setting that determines the processing order of the rules in a rule group. DNS Firewall processes the rules in a rule group by order of priority, starting from the lowest setting.

$sel:action:ListFirewallRules', listFirewallRules_action - Optional additional filter for the rules to retrieve.

The action that DNS Firewall should take on a DNS query when it matches one of the domains in the rule's domain list:

ALLOW - Permit the request to go through.

ALERT - Permit the request to go through but send an alert to the logs.

BLOCK - Disallow the request. If this is specified, additional handling details are provided in the rule's BlockResponse setting.

$sel:nextToken:ListFirewallRules', listFirewallRules_nextToken - For the first call to this list request, omit this value.

When you request a list of objects, Resolver returns at most the number of objects specified in MaxResults. If more objects are available for retrieval, Resolver returns a NextToken value in the response. To retrieve the next batch of objects, use the token that was returned for the prior request in your next request.

$sel:maxResults:ListFirewallRules', listFirewallRules_maxResults - The maximum number of objects that you want Resolver to return for this request. If more objects are available, in the response, Resolver provides a NextToken value that you can use in a subsequent call to get the next batch of objects.

If you don't specify a value for MaxResults, Resolver returns up to 100 objects.

$sel:firewallRuleGroupId:ListFirewallRules', listFirewallRules_firewallRuleGroupId - The unique identifier of the firewall rule group that you want to retrieve the rules for.

Request Lenses

listFirewallRules_priority :: Lens' ListFirewallRules (Maybe Int)

Optional additional filter for the rules to retrieve.

The setting that determines the processing order of the rules in a rule group. DNS Firewall processes the rules in a rule group by order of priority, starting from the lowest setting.

listFirewallRules_action :: Lens' ListFirewallRules (Maybe Action)

Optional additional filter for the rules to retrieve.

The action that DNS Firewall should take on a DNS query when it matches one of the domains in the rule's domain list:

ALLOW - Permit the request to go through.

ALERT - Permit the request to go through but send an alert to the logs.

BLOCK - Disallow the request. If this is specified, additional handling details are provided in the rule's BlockResponse setting.

listFirewallRules_nextToken :: Lens' ListFirewallRules (Maybe Text)

For the first call to this list request, omit this value.

When you request a list of objects, Resolver returns at most the number of objects specified in MaxResults. If more objects are available for retrieval, Resolver returns a NextToken value in the response. To retrieve the next batch of objects, use the token that was returned for the prior request in your next request.

listFirewallRules_maxResults :: Lens' ListFirewallRules (Maybe Natural)

The maximum number of objects that you want Resolver to return for this request. If more objects are available, in the response, Resolver provides a NextToken value that you can use in a subsequent call to get the next batch of objects.

If you don't specify a value for MaxResults, Resolver returns up to 100 objects.

listFirewallRules_firewallRuleGroupId :: Lens' ListFirewallRules Text

The unique identifier of the firewall rule group that you want to retrieve the rules for.

Destructuring the Response

data ListFirewallRulesResponse

See: newListFirewallRulesResponse smart constructor.

Constructors

ListFirewallRulesResponse'

Fields

firewallRules :: Maybe [FirewallRule]

A list of the rules that you have defined.

This might be a partial list of the firewall rules that you've defined. For information, see MaxResults.

nextToken :: Maybe Text

If objects are still available for retrieval, Resolver returns this token in the response. To retrieve the next batch of objects, provide this token in your next request.

httpStatus :: Int

The response's http status code.

newListFirewallRulesResponse

Arguments

:: Int
$sel:httpStatus:ListFirewallRulesResponse'

-> ListFirewallRulesResponse

Create a value of ListFirewallRulesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:firewallRules:ListFirewallRulesResponse', listFirewallRulesResponse_firewallRules - A list of the rules that you have defined.

This might be a partial list of the firewall rules that you've defined. For information, see MaxResults.

$sel:nextToken:ListFirewallRules', listFirewallRulesResponse_nextToken - If objects are still available for retrieval, Resolver returns this token in the response. To retrieve the next batch of objects, provide this token in your next request.

$sel:httpStatus:ListFirewallRulesResponse', listFirewallRulesResponse_httpStatus - The response's http status code.

Response Lenses

listFirewallRulesResponse_firewallRules :: Lens' ListFirewallRulesResponse (Maybe [FirewallRule])

A list of the rules that you have defined.

This might be a partial list of the firewall rules that you've defined. For information, see MaxResults.

listFirewallRulesResponse_nextToken :: Lens' ListFirewallRulesResponse (Maybe Text)

If objects are still available for retrieval, Resolver returns this token in the response. To retrieve the next batch of objects, provide this token in your next request.

listFirewallRulesResponse_httpStatus :: Lens' ListFirewallRulesResponse Int

The response's http status code.