Network.AWS.Route53Resolver.ListResolverDnssecConfigs

Description

Lists the configurations for DNSSEC validation that are associated with the current Amazon Web Services account.

This operation returns paginated results.

Creating a Request

data ListResolverDnssecConfigs

See: newListResolverDnssecConfigs smart constructor.

Constructors

ListResolverDnssecConfigs'

Fields

filters :: Maybe [Filter]

An optional specification to return a subset of objects.

nextToken :: Maybe Text

(Optional) If the current Amazon Web Services account has more than MaxResults DNSSEC configurations, use NextToken to get the second and subsequent pages of results.

For the first ListResolverDnssecConfigs request, omit this value.

For the second and subsequent requests, get the value of NextToken from the previous response and specify that value for NextToken in the request.

maxResults :: Maybe Natural

Optional: An integer that specifies the maximum number of DNSSEC configuration results that you want Amazon Route 53 to return. If you don't specify a value for MaxResults, Route 53 returns up to 100 configuration per page.

newListResolverDnssecConfigs :: ListResolverDnssecConfigs

Create a value of ListResolverDnssecConfigs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListResolverDnssecConfigs', listResolverDnssecConfigs_filters - An optional specification to return a subset of objects.

$sel:nextToken:ListResolverDnssecConfigs', listResolverDnssecConfigs_nextToken - (Optional) If the current Amazon Web Services account has more than MaxResults DNSSEC configurations, use NextToken to get the second and subsequent pages of results.

For the first ListResolverDnssecConfigs request, omit this value.

For the second and subsequent requests, get the value of NextToken from the previous response and specify that value for NextToken in the request.

$sel:maxResults:ListResolverDnssecConfigs', listResolverDnssecConfigs_maxResults - Optional: An integer that specifies the maximum number of DNSSEC configuration results that you want Amazon Route 53 to return. If you don't specify a value for MaxResults, Route 53 returns up to 100 configuration per page.

Request Lenses

listResolverDnssecConfigs_filters :: Lens' ListResolverDnssecConfigs (Maybe [Filter])

An optional specification to return a subset of objects.

listResolverDnssecConfigs_nextToken :: Lens' ListResolverDnssecConfigs (Maybe Text)

(Optional) If the current Amazon Web Services account has more than MaxResults DNSSEC configurations, use NextToken to get the second and subsequent pages of results.

For the first ListResolverDnssecConfigs request, omit this value.

For the second and subsequent requests, get the value of NextToken from the previous response and specify that value for NextToken in the request.

listResolverDnssecConfigs_maxResults :: Lens' ListResolverDnssecConfigs (Maybe Natural)

Optional: An integer that specifies the maximum number of DNSSEC configuration results that you want Amazon Route 53 to return. If you don't specify a value for MaxResults, Route 53 returns up to 100 configuration per page.

Destructuring the Response

data ListResolverDnssecConfigsResponse

See: newListResolverDnssecConfigsResponse smart constructor.

Constructors

ListResolverDnssecConfigsResponse'

Fields

nextToken :: Maybe Text

If a response includes the last of the DNSSEC configurations that are associated with the current Amazon Web Services account, NextToken doesn't appear in the response.

If a response doesn't include the last of the configurations, you can get more configurations by submitting another ListResolverDnssecConfigs request. Get the value of NextToken that Amazon Route 53 returned in the previous response and include it in NextToken in the next request.

resolverDnssecConfigs :: Maybe [ResolverDnssecConfig]

An array that contains one ResolverDnssecConfig element for each configuration for DNSSEC validation that is associated with the current Amazon Web Services account.

httpStatus :: Int

The response's http status code.

newListResolverDnssecConfigsResponse

Arguments

:: Int
$sel:httpStatus:ListResolverDnssecConfigsResponse'

-> ListResolverDnssecConfigsResponse

Create a value of ListResolverDnssecConfigsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListResolverDnssecConfigs', listResolverDnssecConfigsResponse_nextToken - If a response includes the last of the DNSSEC configurations that are associated with the current Amazon Web Services account, NextToken doesn't appear in the response.

If a response doesn't include the last of the configurations, you can get more configurations by submitting another ListResolverDnssecConfigs request. Get the value of NextToken that Amazon Route 53 returned in the previous response and include it in NextToken in the next request.

$sel:resolverDnssecConfigs:ListResolverDnssecConfigsResponse', listResolverDnssecConfigsResponse_resolverDnssecConfigs - An array that contains one ResolverDnssecConfig element for each configuration for DNSSEC validation that is associated with the current Amazon Web Services account.

$sel:httpStatus:ListResolverDnssecConfigsResponse', listResolverDnssecConfigsResponse_httpStatus - The response's http status code.

Response Lenses

listResolverDnssecConfigsResponse_nextToken :: Lens' ListResolverDnssecConfigsResponse (Maybe Text)

If a response includes the last of the DNSSEC configurations that are associated with the current Amazon Web Services account, NextToken doesn't appear in the response.

If a response doesn't include the last of the configurations, you can get more configurations by submitting another ListResolverDnssecConfigs request. Get the value of NextToken that Amazon Route 53 returned in the previous response and include it in NextToken in the next request.

listResolverDnssecConfigsResponse_resolverDnssecConfigs :: Lens' ListResolverDnssecConfigsResponse (Maybe [ResolverDnssecConfig])

An array that contains one ResolverDnssecConfig element for each configuration for DNSSEC validation that is associated with the current Amazon Web Services account.

listResolverDnssecConfigsResponse_httpStatus :: Lens' ListResolverDnssecConfigsResponse Int

The response's http status code.