Network.AWS.Route53Resolver.ListResolverRuleAssociations

Description

Lists the associations that were created between Resolver rules and VPCs using the current Amazon Web Services account.

This operation returns paginated results.

Creating a Request

data ListResolverRuleAssociations

See: newListResolverRuleAssociations smart constructor.

Constructors

ListResolverRuleAssociations'

Fields

filters :: Maybe [Filter]

An optional specification to return a subset of Resolver rules, such as Resolver rules that are associated with the same VPC ID.

If you submit a second or subsequent ListResolverRuleAssociations request and specify the NextToken parameter, you must use the same values for Filters, if any, as in the previous request.

nextToken :: Maybe Text

For the first ListResolverRuleAssociation request, omit this value.

If you have more than MaxResults rule associations, you can submit another ListResolverRuleAssociation request to get the next group of rule associations. In the next request, specify the value of NextToken from the previous response.

maxResults :: Maybe Natural

The maximum number of rule associations that you want to return in the response to a ListResolverRuleAssociations request. If you don't specify a value for MaxResults, Resolver returns up to 100 rule associations.

newListResolverRuleAssociations :: ListResolverRuleAssociations

Create a value of ListResolverRuleAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListResolverRuleAssociations', listResolverRuleAssociations_filters - An optional specification to return a subset of Resolver rules, such as Resolver rules that are associated with the same VPC ID.

If you submit a second or subsequent ListResolverRuleAssociations request and specify the NextToken parameter, you must use the same values for Filters, if any, as in the previous request.

$sel:nextToken:ListResolverRuleAssociations', listResolverRuleAssociations_nextToken - For the first ListResolverRuleAssociation request, omit this value.

If you have more than MaxResults rule associations, you can submit another ListResolverRuleAssociation request to get the next group of rule associations. In the next request, specify the value of NextToken from the previous response.

$sel:maxResults:ListResolverRuleAssociations', listResolverRuleAssociations_maxResults - The maximum number of rule associations that you want to return in the response to a ListResolverRuleAssociations request. If you don't specify a value for MaxResults, Resolver returns up to 100 rule associations.

Request Lenses

listResolverRuleAssociations_filters :: Lens' ListResolverRuleAssociations (Maybe [Filter])

An optional specification to return a subset of Resolver rules, such as Resolver rules that are associated with the same VPC ID.

If you submit a second or subsequent ListResolverRuleAssociations request and specify the NextToken parameter, you must use the same values for Filters, if any, as in the previous request.

listResolverRuleAssociations_nextToken :: Lens' ListResolverRuleAssociations (Maybe Text)

For the first ListResolverRuleAssociation request, omit this value.

If you have more than MaxResults rule associations, you can submit another ListResolverRuleAssociation request to get the next group of rule associations. In the next request, specify the value of NextToken from the previous response.

listResolverRuleAssociations_maxResults :: Lens' ListResolverRuleAssociations (Maybe Natural)

The maximum number of rule associations that you want to return in the response to a ListResolverRuleAssociations request. If you don't specify a value for MaxResults, Resolver returns up to 100 rule associations.

Destructuring the Response

data ListResolverRuleAssociationsResponse

See: newListResolverRuleAssociationsResponse smart constructor.

Constructors

ListResolverRuleAssociationsResponse'

Fields

resolverRuleAssociations :: Maybe [ResolverRuleAssociation]

The associations that were created between Resolver rules and VPCs using the current Amazon Web Services account, and that match the specified filters, if any.

nextToken :: Maybe Text

If more than MaxResults rule associations match the specified criteria, you can submit another ListResolverRuleAssociation request to get the next group of results. In the next request, specify the value of NextToken from the previous response.

maxResults :: Maybe Natural

The value that you specified for MaxResults in the request.

httpStatus :: Int

The response's http status code.

newListResolverRuleAssociationsResponse

Arguments

:: Int
$sel:httpStatus:ListResolverRuleAssociationsResponse'

-> ListResolverRuleAssociationsResponse

Create a value of ListResolverRuleAssociationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolverRuleAssociations:ListResolverRuleAssociationsResponse', listResolverRuleAssociationsResponse_resolverRuleAssociations - The associations that were created between Resolver rules and VPCs using the current Amazon Web Services account, and that match the specified filters, if any.

$sel:nextToken:ListResolverRuleAssociations', listResolverRuleAssociationsResponse_nextToken - If more than MaxResults rule associations match the specified criteria, you can submit another ListResolverRuleAssociation request to get the next group of results. In the next request, specify the value of NextToken from the previous response.

$sel:maxResults:ListResolverRuleAssociations', listResolverRuleAssociationsResponse_maxResults - The value that you specified for MaxResults in the request.

$sel:httpStatus:ListResolverRuleAssociationsResponse', listResolverRuleAssociationsResponse_httpStatus - The response's http status code.

Response Lenses

listResolverRuleAssociationsResponse_resolverRuleAssociations :: Lens' ListResolverRuleAssociationsResponse (Maybe [ResolverRuleAssociation])

The associations that were created between Resolver rules and VPCs using the current Amazon Web Services account, and that match the specified filters, if any.

listResolverRuleAssociationsResponse_nextToken :: Lens' ListResolverRuleAssociationsResponse (Maybe Text)

If more than MaxResults rule associations match the specified criteria, you can submit another ListResolverRuleAssociation request to get the next group of results. In the next request, specify the value of NextToken from the previous response.

listResolverRuleAssociationsResponse_maxResults :: Lens' ListResolverRuleAssociationsResponse (Maybe Natural)

The value that you specified for MaxResults in the request.

listResolverRuleAssociationsResponse_httpStatus :: Lens' ListResolverRuleAssociationsResponse Int

The response's http status code.