Network.AWS.Route53Resolver.PutResolverQueryLogConfigPolicy

Description

Specifies an Amazon Web Services account that you want to share a query logging configuration with, the query logging configuration that you want to share, and the operations that you want the account to be able to perform on the configuration.

Creating a Request

data PutResolverQueryLogConfigPolicy

See: newPutResolverQueryLogConfigPolicy smart constructor.

Constructors

PutResolverQueryLogConfigPolicy'

Fields

arn :: Text

The Amazon Resource Name (ARN) of the account that you want to share rules with.

resolverQueryLogConfigPolicy :: Text

An Identity and Access Management policy statement that lists the query logging configurations that you want to share with another Amazon Web Services account and the operations that you want the account to be able to perform. You can specify the following operations in the Actions section of the statement:

route53resolver:AssociateResolverQueryLogConfig

route53resolver:DisassociateResolverQueryLogConfig

route53resolver:ListResolverQueryLogConfigAssociations

route53resolver:ListResolverQueryLogConfigs

In the Resource section of the statement, you specify the ARNs for the query logging configurations that you want to share with the account that you specified in Arn.

newPutResolverQueryLogConfigPolicy

Arguments

:: Text
$sel:arn:PutResolverQueryLogConfigPolicy'

-> Text
$sel:resolverQueryLogConfigPolicy:PutResolverQueryLogConfigPolicy'

-> PutResolverQueryLogConfigPolicy

Create a value of PutResolverQueryLogConfigPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:PutResolverQueryLogConfigPolicy', putResolverQueryLogConfigPolicy_arn - The Amazon Resource Name (ARN) of the account that you want to share rules with.

$sel:resolverQueryLogConfigPolicy:PutResolverQueryLogConfigPolicy', putResolverQueryLogConfigPolicy_resolverQueryLogConfigPolicy - An Identity and Access Management policy statement that lists the query logging configurations that you want to share with another Amazon Web Services account and the operations that you want the account to be able to perform. You can specify the following operations in the Actions section of the statement:

route53resolver:AssociateResolverQueryLogConfig

route53resolver:DisassociateResolverQueryLogConfig

route53resolver:ListResolverQueryLogConfigAssociations

route53resolver:ListResolverQueryLogConfigs

In the Resource section of the statement, you specify the ARNs for the query logging configurations that you want to share with the account that you specified in Arn.

Request Lenses

putResolverQueryLogConfigPolicy_arn :: Lens' PutResolverQueryLogConfigPolicy Text

The Amazon Resource Name (ARN) of the account that you want to share rules with.

putResolverQueryLogConfigPolicy_resolverQueryLogConfigPolicy :: Lens' PutResolverQueryLogConfigPolicy Text

An Identity and Access Management policy statement that lists the query logging configurations that you want to share with another Amazon Web Services account and the operations that you want the account to be able to perform. You can specify the following operations in the Actions section of the statement:

route53resolver:AssociateResolverQueryLogConfig

route53resolver:DisassociateResolverQueryLogConfig

route53resolver:ListResolverQueryLogConfigAssociations

route53resolver:ListResolverQueryLogConfigs

In the Resource section of the statement, you specify the ARNs for the query logging configurations that you want to share with the account that you specified in Arn.

Destructuring the Response

data PutResolverQueryLogConfigPolicyResponse

The response to a PutResolverQueryLogConfigPolicy request.

See: newPutResolverQueryLogConfigPolicyResponse smart constructor.

Constructors

PutResolverQueryLogConfigPolicyResponse'

Fields

returnValue :: Maybe Bool

Whether the PutResolverQueryLogConfigPolicy request was successful.

httpStatus :: Int

The response's http status code.

newPutResolverQueryLogConfigPolicyResponse

Arguments

:: Int
$sel:httpStatus:PutResolverQueryLogConfigPolicyResponse'

-> PutResolverQueryLogConfigPolicyResponse

Create a value of PutResolverQueryLogConfigPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:returnValue:PutResolverQueryLogConfigPolicyResponse', putResolverQueryLogConfigPolicyResponse_returnValue - Whether the PutResolverQueryLogConfigPolicy request was successful.

$sel:httpStatus:PutResolverQueryLogConfigPolicyResponse', putResolverQueryLogConfigPolicyResponse_httpStatus - The response's http status code.

Response Lenses

putResolverQueryLogConfigPolicyResponse_returnValue :: Lens' PutResolverQueryLogConfigPolicyResponse (Maybe Bool)

Whether the PutResolverQueryLogConfigPolicy request was successful.

putResolverQueryLogConfigPolicyResponse_httpStatus :: Lens' PutResolverQueryLogConfigPolicyResponse Int

The response's http status code.