Network.AWS.Route53Resolver.Types.FirewallConfig

Description

Documentation

data FirewallConfig

Configuration of the firewall behavior provided by DNS Firewall for a single VPC from Amazon Virtual Private Cloud (Amazon VPC).

See: newFirewallConfig smart constructor.

Constructors

FirewallConfig'

Fields

resourceId :: Maybe Text

The ID of the VPC that this firewall configuration applies to.

ownerId :: Maybe Text

The Amazon Web Services account ID of the owner of the VPC that this firewall configuration applies to.

id :: Maybe Text

The ID of the firewall configuration.

firewallFailOpen :: Maybe FirewallFailOpenStatus

Determines how DNS Firewall operates during failures, for example when all traffic that is sent to DNS Firewall fails to receive a reply.

By default, fail open is disabled, which means the failure mode is closed. This approach favors security over availability. DNS Firewall returns a failure error when it is unable to properly evaluate a query.

If you enable this option, the failure mode is open. This approach favors availability over security. DNS Firewall allows queries to proceed if it is unable to properly evaluate them.

This behavior is only enforced for VPCs that have at least one DNS Firewall rule group association.

newFirewallConfig :: FirewallConfig

Create a value of FirewallConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:FirewallConfig', firewallConfig_resourceId - The ID of the VPC that this firewall configuration applies to.

$sel:ownerId:FirewallConfig', firewallConfig_ownerId - The Amazon Web Services account ID of the owner of the VPC that this firewall configuration applies to.

$sel:id:FirewallConfig', firewallConfig_id - The ID of the firewall configuration.

$sel:firewallFailOpen:FirewallConfig', firewallConfig_firewallFailOpen - Determines how DNS Firewall operates during failures, for example when all traffic that is sent to DNS Firewall fails to receive a reply.

By default, fail open is disabled, which means the failure mode is closed. This approach favors security over availability. DNS Firewall returns a failure error when it is unable to properly evaluate a query.

If you enable this option, the failure mode is open. This approach favors availability over security. DNS Firewall allows queries to proceed if it is unable to properly evaluate them.

This behavior is only enforced for VPCs that have at least one DNS Firewall rule group association.

firewallConfig_resourceId :: Lens' FirewallConfig (Maybe Text)

The ID of the VPC that this firewall configuration applies to.

firewallConfig_ownerId :: Lens' FirewallConfig (Maybe Text)

The Amazon Web Services account ID of the owner of the VPC that this firewall configuration applies to.

firewallConfig_id :: Lens' FirewallConfig (Maybe Text)

The ID of the firewall configuration.

firewallConfig_firewallFailOpen :: Lens' FirewallConfig (Maybe FirewallFailOpenStatus)

Determines how DNS Firewall operates during failures, for example when all traffic that is sent to DNS Firewall fails to receive a reply.

By default, fail open is disabled, which means the failure mode is closed. This approach favors security over availability. DNS Firewall returns a failure error when it is unable to properly evaluate a query.

If you enable this option, the failure mode is open. This approach favors availability over security. DNS Firewall allows queries to proceed if it is unable to properly evaluate them.

This behavior is only enforced for VPCs that have at least one DNS Firewall rule group association.