Network.AWS.Route53Resolver.Types.FirewallRule

Description

Documentation

data FirewallRule

A single firewall rule in a rule group.

See: newFirewallRule smart constructor.

Constructors

FirewallRule'

Fields

creationTime :: Maybe Text

The date and time that the rule was created, in Unix time format and Coordinated Universal Time (UTC).

blockOverrideDnsType :: Maybe BlockOverrideDnsType

The DNS record's type. This determines the format of the record value that you provided in BlockOverrideDomain. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

firewallRuleGroupId :: Maybe Text

The unique identifier of the firewall rule group of the rule.

priority :: Maybe Int

The priority of the rule in the rule group. This value must be unique within the rule group. DNS Firewall processes the rules in a rule group by order of priority, starting from the lowest setting.

blockResponse :: Maybe BlockResponse

The way that you want DNS Firewall to block the request. Used for the rule action setting BLOCK.

NODATA - Respond indicating that the query was successful, but no response is available for it.

NXDOMAIN - Respond indicating that the domain name that's in the query doesn't exist.

OVERRIDE - Provide a custom override in the response. This option requires custom handling details in the rule's BlockOverride* settings.

creatorRequestId :: Maybe Text

A unique string defined by you to identify the request. This allows you to retry failed requests without the risk of executing the operation twice. This can be any unique string, for example, a timestamp.

modificationTime :: Maybe Text

The date and time that the rule was last modified, in Unix time format and Coordinated Universal Time (UTC).

action :: Maybe Action

The action that DNS Firewall should take on a DNS query when it matches one of the domains in the rule's domain list:

ALLOW - Permit the request to go through.

ALERT - Permit the request to go through but send an alert to the logs.

BLOCK - Disallow the request. If this is specified, additional handling details are provided in the rule's BlockResponse setting.

blockOverrideTtl :: Maybe Int

The recommended amount of time, in seconds, for the DNS resolver or web browser to cache the provided override record. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

name :: Maybe Text

The name of the rule.

blockOverrideDomain :: Maybe Text

The custom DNS record to send back in response to the query. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

firewallDomainListId :: Maybe Text

The ID of the domain list that's used in the rule.

newFirewallRule :: FirewallRule

Create a value of FirewallRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:FirewallRule', firewallRule_creationTime - The date and time that the rule was created, in Unix time format and Coordinated Universal Time (UTC).

$sel:blockOverrideDnsType:FirewallRule', firewallRule_blockOverrideDnsType - The DNS record's type. This determines the format of the record value that you provided in BlockOverrideDomain. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

$sel:firewallRuleGroupId:FirewallRule', firewallRule_firewallRuleGroupId - The unique identifier of the firewall rule group of the rule.

$sel:priority:FirewallRule', firewallRule_priority - The priority of the rule in the rule group. This value must be unique within the rule group. DNS Firewall processes the rules in a rule group by order of priority, starting from the lowest setting.

$sel:blockResponse:FirewallRule', firewallRule_blockResponse - The way that you want DNS Firewall to block the request. Used for the rule action setting BLOCK.

NODATA - Respond indicating that the query was successful, but no response is available for it.

NXDOMAIN - Respond indicating that the domain name that's in the query doesn't exist.

OVERRIDE - Provide a custom override in the response. This option requires custom handling details in the rule's BlockOverride* settings.

$sel:creatorRequestId:FirewallRule', firewallRule_creatorRequestId - A unique string defined by you to identify the request. This allows you to retry failed requests without the risk of executing the operation twice. This can be any unique string, for example, a timestamp.

$sel:modificationTime:FirewallRule', firewallRule_modificationTime - The date and time that the rule was last modified, in Unix time format and Coordinated Universal Time (UTC).

$sel:action:FirewallRule', firewallRule_action - The action that DNS Firewall should take on a DNS query when it matches one of the domains in the rule's domain list:

ALLOW - Permit the request to go through.

ALERT - Permit the request to go through but send an alert to the logs.

BLOCK - Disallow the request. If this is specified, additional handling details are provided in the rule's BlockResponse setting.

$sel:blockOverrideTtl:FirewallRule', firewallRule_blockOverrideTtl - The recommended amount of time, in seconds, for the DNS resolver or web browser to cache the provided override record. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

$sel:name:FirewallRule', firewallRule_name - The name of the rule.

$sel:blockOverrideDomain:FirewallRule', firewallRule_blockOverrideDomain - The custom DNS record to send back in response to the query. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

$sel:firewallDomainListId:FirewallRule', firewallRule_firewallDomainListId - The ID of the domain list that's used in the rule.

firewallRule_creationTime :: Lens' FirewallRule (Maybe Text)

The date and time that the rule was created, in Unix time format and Coordinated Universal Time (UTC).

firewallRule_blockOverrideDnsType :: Lens' FirewallRule (Maybe BlockOverrideDnsType)

The DNS record's type. This determines the format of the record value that you provided in BlockOverrideDomain. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

firewallRule_firewallRuleGroupId :: Lens' FirewallRule (Maybe Text)

The unique identifier of the firewall rule group of the rule.

firewallRule_priority :: Lens' FirewallRule (Maybe Int)

The priority of the rule in the rule group. This value must be unique within the rule group. DNS Firewall processes the rules in a rule group by order of priority, starting from the lowest setting.

firewallRule_blockResponse :: Lens' FirewallRule (Maybe BlockResponse)

The way that you want DNS Firewall to block the request. Used for the rule action setting BLOCK.

NODATA - Respond indicating that the query was successful, but no response is available for it.

NXDOMAIN - Respond indicating that the domain name that's in the query doesn't exist.

OVERRIDE - Provide a custom override in the response. This option requires custom handling details in the rule's BlockOverride* settings.

firewallRule_creatorRequestId :: Lens' FirewallRule (Maybe Text)

A unique string defined by you to identify the request. This allows you to retry failed requests without the risk of executing the operation twice. This can be any unique string, for example, a timestamp.

firewallRule_modificationTime :: Lens' FirewallRule (Maybe Text)

The date and time that the rule was last modified, in Unix time format and Coordinated Universal Time (UTC).

firewallRule_action :: Lens' FirewallRule (Maybe Action)

The action that DNS Firewall should take on a DNS query when it matches one of the domains in the rule's domain list:

ALLOW - Permit the request to go through.

ALERT - Permit the request to go through but send an alert to the logs.

BLOCK - Disallow the request. If this is specified, additional handling details are provided in the rule's BlockResponse setting.

firewallRule_blockOverrideTtl :: Lens' FirewallRule (Maybe Int)

The recommended amount of time, in seconds, for the DNS resolver or web browser to cache the provided override record. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

firewallRule_name :: Lens' FirewallRule (Maybe Text)

The name of the rule.

firewallRule_blockOverrideDomain :: Lens' FirewallRule (Maybe Text)

The custom DNS record to send back in response to the query. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

firewallRule_firewallDomainListId :: Lens' FirewallRule (Maybe Text)

The ID of the domain list that's used in the rule.