Network.AWS.Route53Resolver.Types.FirewallRuleGroupMetadata

Description

Documentation

data FirewallRuleGroupMetadata

Minimal high-level information for a firewall rule group. The action ListFirewallRuleGroups returns an array of these objects.

To retrieve full information for a firewall rule group, call GetFirewallRuleGroup and ListFirewallRules.

See: newFirewallRuleGroupMetadata smart constructor.

Constructors

FirewallRuleGroupMetadata'

Fields

arn :: Maybe Text

The ARN (Amazon Resource Name) of the rule group.

creatorRequestId :: Maybe Text

A unique string defined by you to identify the request. This allows you to retry failed requests without the risk of running the operation twice. This can be any unique string, for example, a timestamp.

shareStatus :: Maybe ShareStatus

Whether the rule group is shared with other Amazon Web Services accounts, or was shared with the current account by another Amazon Web Services account. Sharing is configured through Resource Access Manager (RAM).

ownerId :: Maybe Text

The Amazon Web Services account ID for the account that created the rule group. When a rule group is shared with your account, this is the account that has shared the rule group with you.

name :: Maybe Text

The name of the rule group.

id :: Maybe Text

The ID of the rule group.

newFirewallRuleGroupMetadata :: FirewallRuleGroupMetadata

Create a value of FirewallRuleGroupMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:FirewallRuleGroupMetadata', firewallRuleGroupMetadata_arn - The ARN (Amazon Resource Name) of the rule group.

$sel:creatorRequestId:FirewallRuleGroupMetadata', firewallRuleGroupMetadata_creatorRequestId - A unique string defined by you to identify the request. This allows you to retry failed requests without the risk of running the operation twice. This can be any unique string, for example, a timestamp.

$sel:shareStatus:FirewallRuleGroupMetadata', firewallRuleGroupMetadata_shareStatus - Whether the rule group is shared with other Amazon Web Services accounts, or was shared with the current account by another Amazon Web Services account. Sharing is configured through Resource Access Manager (RAM).

$sel:ownerId:FirewallRuleGroupMetadata', firewallRuleGroupMetadata_ownerId - The Amazon Web Services account ID for the account that created the rule group. When a rule group is shared with your account, this is the account that has shared the rule group with you.

$sel:name:FirewallRuleGroupMetadata', firewallRuleGroupMetadata_name - The name of the rule group.

$sel:id:FirewallRuleGroupMetadata', firewallRuleGroupMetadata_id - The ID of the rule group.

firewallRuleGroupMetadata_arn :: Lens' FirewallRuleGroupMetadata (Maybe Text)

The ARN (Amazon Resource Name) of the rule group.

firewallRuleGroupMetadata_creatorRequestId :: Lens' FirewallRuleGroupMetadata (Maybe Text)

A unique string defined by you to identify the request. This allows you to retry failed requests without the risk of running the operation twice. This can be any unique string, for example, a timestamp.

firewallRuleGroupMetadata_shareStatus :: Lens' FirewallRuleGroupMetadata (Maybe ShareStatus)

Whether the rule group is shared with other Amazon Web Services accounts, or was shared with the current account by another Amazon Web Services account. Sharing is configured through Resource Access Manager (RAM).

firewallRuleGroupMetadata_ownerId :: Lens' FirewallRuleGroupMetadata (Maybe Text)

The Amazon Web Services account ID for the account that created the rule group. When a rule group is shared with your account, this is the account that has shared the rule group with you.

firewallRuleGroupMetadata_name :: Lens' FirewallRuleGroupMetadata (Maybe Text)

The name of the rule group.

firewallRuleGroupMetadata_id :: Lens' FirewallRuleGroupMetadata (Maybe Text)

The ID of the rule group.