Network.AWS.Route53Resolver.Types.IpAddressRequest

Description

Documentation

data IpAddressRequest

In a CreateResolverEndpoint request, the IP address that DNS queries originate from (for outbound endpoints) or that you forward DNS queries to (for inbound endpoints). IpAddressRequest also includes the ID of the subnet that contains the IP address.

See: newIpAddressRequest smart constructor.

Constructors

IpAddressRequest'

Fields

ip :: Maybe Text

The IP address that you want to use for DNS queries.

subnetId :: Text

The ID of the subnet that contains the IP address.

newIpAddressRequest

Arguments

:: Text
$sel:subnetId:IpAddressRequest'

-> IpAddressRequest

Create a value of IpAddressRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ip:IpAddressRequest', ipAddressRequest_ip - The IP address that you want to use for DNS queries.

$sel:subnetId:IpAddressRequest', ipAddressRequest_subnetId - The ID of the subnet that contains the IP address.

ipAddressRequest_ip :: Lens' IpAddressRequest (Maybe Text)

The IP address that you want to use for DNS queries.

ipAddressRequest_subnetId :: Lens' IpAddressRequest Text

The ID of the subnet that contains the IP address.