Network.AWS.Route53Resolver.Types.IpAddressResponse

Description

Documentation

data IpAddressResponse

In the response to a GetResolverEndpoint request, information about the IP addresses that the Resolver endpoint uses for DNS queries.

See: newIpAddressResponse smart constructor.

Constructors

IpAddressResponse'

Fields

creationTime :: Maybe Text

The date and time that the IP address was created, in Unix time format and Coordinated Universal Time (UTC).

status :: Maybe IpAddressStatus

A status code that gives the current status of the request.

modificationTime :: Maybe Text

The date and time that the IP address was last modified, in Unix time format and Coordinated Universal Time (UTC).

subnetId :: Maybe Text

The ID of one subnet.

ip :: Maybe Text

One IP address that the Resolver endpoint uses for DNS queries.

ipId :: Maybe Text

The ID of one IP address.

statusMessage :: Maybe Text

A message that provides additional information about the status of the request.

newIpAddressResponse :: IpAddressResponse

Create a value of IpAddressResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:IpAddressResponse', ipAddressResponse_creationTime - The date and time that the IP address was created, in Unix time format and Coordinated Universal Time (UTC).

$sel:status:IpAddressResponse', ipAddressResponse_status - A status code that gives the current status of the request.

$sel:modificationTime:IpAddressResponse', ipAddressResponse_modificationTime - The date and time that the IP address was last modified, in Unix time format and Coordinated Universal Time (UTC).

$sel:subnetId:IpAddressResponse', ipAddressResponse_subnetId - The ID of one subnet.

$sel:ip:IpAddressResponse', ipAddressResponse_ip - One IP address that the Resolver endpoint uses for DNS queries.

$sel:ipId:IpAddressResponse', ipAddressResponse_ipId - The ID of one IP address.

$sel:statusMessage:IpAddressResponse', ipAddressResponse_statusMessage - A message that provides additional information about the status of the request.

ipAddressResponse_creationTime :: Lens' IpAddressResponse (Maybe Text)

The date and time that the IP address was created, in Unix time format and Coordinated Universal Time (UTC).

ipAddressResponse_status :: Lens' IpAddressResponse (Maybe IpAddressStatus)

A status code that gives the current status of the request.

ipAddressResponse_modificationTime :: Lens' IpAddressResponse (Maybe Text)

The date and time that the IP address was last modified, in Unix time format and Coordinated Universal Time (UTC).

ipAddressResponse_subnetId :: Lens' IpAddressResponse (Maybe Text)

The ID of one subnet.

ipAddressResponse_ip :: Lens' IpAddressResponse (Maybe Text)

One IP address that the Resolver endpoint uses for DNS queries.

ipAddressResponse_ipId :: Lens' IpAddressResponse (Maybe Text)

The ID of one IP address.

ipAddressResponse_statusMessage :: Lens' IpAddressResponse (Maybe Text)

A message that provides additional information about the status of the request.