Network.AWS.Route53Resolver.Types.ResolverEndpoint

Description

Documentation

data ResolverEndpoint

In the response to a CreateResolverEndpoint, DeleteResolverEndpoint, GetResolverEndpoint, ListResolverEndpoints, or UpdateResolverEndpoint request, a complex type that contains settings for an existing inbound or outbound Resolver endpoint.

See: newResolverEndpoint smart constructor.

Constructors

ResolverEndpoint'

Fields

creationTime :: Maybe Text

The date and time that the endpoint was created, in Unix time format and Coordinated Universal Time (UTC).

status :: Maybe ResolverEndpointStatus

A code that specifies the current status of the Resolver endpoint. Valid values include the following:

CREATING: Resolver is creating and configuring one or more Amazon VPC network interfaces for this endpoint.

OPERATIONAL: The Amazon VPC network interfaces for this endpoint are correctly configured and able to pass inbound or outbound DNS queries between your network and Resolver.

UPDATING: Resolver is associating or disassociating one or more network interfaces with this endpoint.

AUTO_RECOVERING: Resolver is trying to recover one or more of the network interfaces that are associated with this endpoint. During the recovery process, the endpoint functions with limited capacity because of the limit on the number of DNS queries per IP address (per network interface). For the current limit, see Limits on Route 53 Resolver.

ACTION_NEEDED: This endpoint is unhealthy, and Resolver can't automatically recover it. To resolve the problem, we recommend that you check each IP address that you associated with the endpoint. For each IP address that isn't available, add another IP address and then delete the IP address that isn't available. (An endpoint must always include at least two IP addresses.) A status of ACTION_NEEDED can have a variety of causes. Here are two common causes:

One or more of the network interfaces that are associated with the endpoint were deleted using Amazon VPC.

The network interface couldn't be created for some reason that's outside the control of Resolver.

DELETING: Resolver is deleting this endpoint and the associated network interfaces.

securityGroupIds :: Maybe [Text]

The ID of one or more security groups that control access to this VPC. The security group must include one or more inbound rules (for inbound endpoints) or outbound rules (for outbound endpoints). Inbound and outbound rules must allow TCP and UDP access. For inbound access, open port 53. For outbound access, open the port that you're using for DNS queries on your network.

direction :: Maybe ResolverEndpointDirection

Indicates whether the Resolver endpoint allows inbound or outbound DNS queries:

INBOUND: allows DNS queries to your VPC from your network

OUTBOUND: allows DNS queries from your VPC to your network

arn :: Maybe Text

The ARN (Amazon Resource Name) for the Resolver endpoint.

creatorRequestId :: Maybe Text

A unique string that identifies the request that created the Resolver endpoint. The CreatorRequestId allows failed requests to be retried without the risk of running the operation twice.

modificationTime :: Maybe Text

The date and time that the endpoint was last modified, in Unix time format and Coordinated Universal Time (UTC).

ipAddressCount :: Maybe Int

The number of IP addresses that the Resolver endpoint can use for DNS queries.

statusMessage :: Maybe Text

A detailed description of the status of the Resolver endpoint.

name :: Maybe Text

The name that you assigned to the Resolver endpoint when you submitted a CreateResolverEndpoint request.

id :: Maybe Text

The ID of the Resolver endpoint.

hostVPCId :: Maybe Text

The ID of the VPC that you want to create the Resolver endpoint in.

newResolverEndpoint :: ResolverEndpoint

Create a value of ResolverEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ResolverEndpoint', resolverEndpoint_creationTime - The date and time that the endpoint was created, in Unix time format and Coordinated Universal Time (UTC).

$sel:status:ResolverEndpoint', resolverEndpoint_status - A code that specifies the current status of the Resolver endpoint. Valid values include the following:

CREATING: Resolver is creating and configuring one or more Amazon VPC network interfaces for this endpoint.

OPERATIONAL: The Amazon VPC network interfaces for this endpoint are correctly configured and able to pass inbound or outbound DNS queries between your network and Resolver.

UPDATING: Resolver is associating or disassociating one or more network interfaces with this endpoint.

AUTO_RECOVERING: Resolver is trying to recover one or more of the network interfaces that are associated with this endpoint. During the recovery process, the endpoint functions with limited capacity because of the limit on the number of DNS queries per IP address (per network interface). For the current limit, see Limits on Route 53 Resolver.

ACTION_NEEDED: This endpoint is unhealthy, and Resolver can't automatically recover it. To resolve the problem, we recommend that you check each IP address that you associated with the endpoint. For each IP address that isn't available, add another IP address and then delete the IP address that isn't available. (An endpoint must always include at least two IP addresses.) A status of ACTION_NEEDED can have a variety of causes. Here are two common causes:

One or more of the network interfaces that are associated with the endpoint were deleted using Amazon VPC.

The network interface couldn't be created for some reason that's outside the control of Resolver.

DELETING: Resolver is deleting this endpoint and the associated network interfaces.

$sel:securityGroupIds:ResolverEndpoint', resolverEndpoint_securityGroupIds - The ID of one or more security groups that control access to this VPC. The security group must include one or more inbound rules (for inbound endpoints) or outbound rules (for outbound endpoints). Inbound and outbound rules must allow TCP and UDP access. For inbound access, open port 53. For outbound access, open the port that you're using for DNS queries on your network.

$sel:direction:ResolverEndpoint', resolverEndpoint_direction - Indicates whether the Resolver endpoint allows inbound or outbound DNS queries:

INBOUND: allows DNS queries to your VPC from your network

OUTBOUND: allows DNS queries from your VPC to your network

$sel:arn:ResolverEndpoint', resolverEndpoint_arn - The ARN (Amazon Resource Name) for the Resolver endpoint.

$sel:creatorRequestId:ResolverEndpoint', resolverEndpoint_creatorRequestId - A unique string that identifies the request that created the Resolver endpoint. The CreatorRequestId allows failed requests to be retried without the risk of running the operation twice.

$sel:modificationTime:ResolverEndpoint', resolverEndpoint_modificationTime - The date and time that the endpoint was last modified, in Unix time format and Coordinated Universal Time (UTC).

$sel:ipAddressCount:ResolverEndpoint', resolverEndpoint_ipAddressCount - The number of IP addresses that the Resolver endpoint can use for DNS queries.

$sel:statusMessage:ResolverEndpoint', resolverEndpoint_statusMessage - A detailed description of the status of the Resolver endpoint.

$sel:name:ResolverEndpoint', resolverEndpoint_name - The name that you assigned to the Resolver endpoint when you submitted a CreateResolverEndpoint request.

$sel:id:ResolverEndpoint', resolverEndpoint_id - The ID of the Resolver endpoint.

$sel:hostVPCId:ResolverEndpoint', resolverEndpoint_hostVPCId - The ID of the VPC that you want to create the Resolver endpoint in.

resolverEndpoint_creationTime :: Lens' ResolverEndpoint (Maybe Text)

The date and time that the endpoint was created, in Unix time format and Coordinated Universal Time (UTC).

resolverEndpoint_status :: Lens' ResolverEndpoint (Maybe ResolverEndpointStatus)

A code that specifies the current status of the Resolver endpoint. Valid values include the following:

CREATING: Resolver is creating and configuring one or more Amazon VPC network interfaces for this endpoint.

OPERATIONAL: The Amazon VPC network interfaces for this endpoint are correctly configured and able to pass inbound or outbound DNS queries between your network and Resolver.

UPDATING: Resolver is associating or disassociating one or more network interfaces with this endpoint.

AUTO_RECOVERING: Resolver is trying to recover one or more of the network interfaces that are associated with this endpoint. During the recovery process, the endpoint functions with limited capacity because of the limit on the number of DNS queries per IP address (per network interface). For the current limit, see Limits on Route 53 Resolver.

ACTION_NEEDED: This endpoint is unhealthy, and Resolver can't automatically recover it. To resolve the problem, we recommend that you check each IP address that you associated with the endpoint. For each IP address that isn't available, add another IP address and then delete the IP address that isn't available. (An endpoint must always include at least two IP addresses.) A status of ACTION_NEEDED can have a variety of causes. Here are two common causes:

One or more of the network interfaces that are associated with the endpoint were deleted using Amazon VPC.

The network interface couldn't be created for some reason that's outside the control of Resolver.

DELETING: Resolver is deleting this endpoint and the associated network interfaces.

resolverEndpoint_securityGroupIds :: Lens' ResolverEndpoint (Maybe [Text])

The ID of one or more security groups that control access to this VPC. The security group must include one or more inbound rules (for inbound endpoints) or outbound rules (for outbound endpoints). Inbound and outbound rules must allow TCP and UDP access. For inbound access, open port 53. For outbound access, open the port that you're using for DNS queries on your network.

resolverEndpoint_direction :: Lens' ResolverEndpoint (Maybe ResolverEndpointDirection)

Indicates whether the Resolver endpoint allows inbound or outbound DNS queries:

INBOUND: allows DNS queries to your VPC from your network

OUTBOUND: allows DNS queries from your VPC to your network

resolverEndpoint_arn :: Lens' ResolverEndpoint (Maybe Text)

The ARN (Amazon Resource Name) for the Resolver endpoint.

resolverEndpoint_creatorRequestId :: Lens' ResolverEndpoint (Maybe Text)

A unique string that identifies the request that created the Resolver endpoint. The CreatorRequestId allows failed requests to be retried without the risk of running the operation twice.

resolverEndpoint_modificationTime :: Lens' ResolverEndpoint (Maybe Text)

The date and time that the endpoint was last modified, in Unix time format and Coordinated Universal Time (UTC).

resolverEndpoint_ipAddressCount :: Lens' ResolverEndpoint (Maybe Int)

The number of IP addresses that the Resolver endpoint can use for DNS queries.

resolverEndpoint_statusMessage :: Lens' ResolverEndpoint (Maybe Text)

A detailed description of the status of the Resolver endpoint.

resolverEndpoint_name :: Lens' ResolverEndpoint (Maybe Text)

The name that you assigned to the Resolver endpoint when you submitted a CreateResolverEndpoint request.

resolverEndpoint_id :: Lens' ResolverEndpoint (Maybe Text)

The ID of the Resolver endpoint.

resolverEndpoint_hostVPCId :: Lens' ResolverEndpoint (Maybe Text)

The ID of the VPC that you want to create the Resolver endpoint in.