Network.AWS.Route53Resolver.Types.ResolverQueryLogConfigAssociation

Description

Documentation

data ResolverQueryLogConfigAssociation

In the response to an AssociateResolverQueryLogConfig, DisassociateResolverQueryLogConfig, GetResolverQueryLogConfigAssociation, or ListResolverQueryLogConfigAssociations, request, a complex type that contains settings for a specified association between an Amazon VPC and a query logging configuration.

See: newResolverQueryLogConfigAssociation smart constructor.

Constructors

ResolverQueryLogConfigAssociation'

Fields

creationTime :: Maybe Text

The date and time that the VPC was associated with the query logging configuration, in Unix time format and Coordinated Universal Time (UTC).

status :: Maybe ResolverQueryLogConfigAssociationStatus

The status of the specified query logging association. Valid values include the following:

CREATING: Resolver is creating an association between an Amazon VPC and a query logging configuration.

CREATED: The association between an Amazon VPC and a query logging configuration was successfully created. Resolver is logging queries that originate in the specified VPC.

DELETING: Resolver is deleting this query logging association.

FAILED: Resolver either couldn't create or couldn't delete the query logging association.

resolverQueryLogConfigId :: Maybe Text

The ID of the query logging configuration that a VPC is associated with.

resourceId :: Maybe Text

The ID of the Amazon VPC that is associated with the query logging configuration.

error :: Maybe ResolverQueryLogConfigAssociationError

If the value of Status is FAILED, the value of Error indicates the cause:

DESTINATION_NOT_FOUND: The specified destination (for example, an Amazon S3 bucket) was deleted.

ACCESS_DENIED: Permissions don't allow sending logs to the destination.

If the value of Status is a value other than FAILED, Error is null.

id :: Maybe Text

The ID of the query logging association.

errorMessage :: Maybe Text

Contains additional information about the error. If the value or Error is null, the value of ErrorMessage also is null.

newResolverQueryLogConfigAssociation :: ResolverQueryLogConfigAssociation

Create a value of ResolverQueryLogConfigAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ResolverQueryLogConfigAssociation', resolverQueryLogConfigAssociation_creationTime - The date and time that the VPC was associated with the query logging configuration, in Unix time format and Coordinated Universal Time (UTC).

$sel:status:ResolverQueryLogConfigAssociation', resolverQueryLogConfigAssociation_status - The status of the specified query logging association. Valid values include the following:

CREATING: Resolver is creating an association between an Amazon VPC and a query logging configuration.

CREATED: The association between an Amazon VPC and a query logging configuration was successfully created. Resolver is logging queries that originate in the specified VPC.

DELETING: Resolver is deleting this query logging association.

FAILED: Resolver either couldn't create or couldn't delete the query logging association.

$sel:resolverQueryLogConfigId:ResolverQueryLogConfigAssociation', resolverQueryLogConfigAssociation_resolverQueryLogConfigId - The ID of the query logging configuration that a VPC is associated with.

$sel:resourceId:ResolverQueryLogConfigAssociation', resolverQueryLogConfigAssociation_resourceId - The ID of the Amazon VPC that is associated with the query logging configuration.

$sel:error:ResolverQueryLogConfigAssociation', resolverQueryLogConfigAssociation_error - If the value of Status is FAILED, the value of Error indicates the cause:

DESTINATION_NOT_FOUND: The specified destination (for example, an Amazon S3 bucket) was deleted.

ACCESS_DENIED: Permissions don't allow sending logs to the destination.

If the value of Status is a value other than FAILED, Error is null.

$sel:id:ResolverQueryLogConfigAssociation', resolverQueryLogConfigAssociation_id - The ID of the query logging association.

$sel:errorMessage:ResolverQueryLogConfigAssociation', resolverQueryLogConfigAssociation_errorMessage - Contains additional information about the error. If the value or Error is null, the value of ErrorMessage also is null.

resolverQueryLogConfigAssociation_creationTime :: Lens' ResolverQueryLogConfigAssociation (Maybe Text)

The date and time that the VPC was associated with the query logging configuration, in Unix time format and Coordinated Universal Time (UTC).

resolverQueryLogConfigAssociation_status :: Lens' ResolverQueryLogConfigAssociation (Maybe ResolverQueryLogConfigAssociationStatus)

The status of the specified query logging association. Valid values include the following:

CREATING: Resolver is creating an association between an Amazon VPC and a query logging configuration.

CREATED: The association between an Amazon VPC and a query logging configuration was successfully created. Resolver is logging queries that originate in the specified VPC.

DELETING: Resolver is deleting this query logging association.

FAILED: Resolver either couldn't create or couldn't delete the query logging association.

resolverQueryLogConfigAssociation_resolverQueryLogConfigId :: Lens' ResolverQueryLogConfigAssociation (Maybe Text)

The ID of the query logging configuration that a VPC is associated with.

resolverQueryLogConfigAssociation_resourceId :: Lens' ResolverQueryLogConfigAssociation (Maybe Text)

The ID of the Amazon VPC that is associated with the query logging configuration.

resolverQueryLogConfigAssociation_error :: Lens' ResolverQueryLogConfigAssociation (Maybe ResolverQueryLogConfigAssociationError)

If the value of Status is FAILED, the value of Error indicates the cause:

DESTINATION_NOT_FOUND: The specified destination (for example, an Amazon S3 bucket) was deleted.

ACCESS_DENIED: Permissions don't allow sending logs to the destination.

If the value of Status is a value other than FAILED, Error is null.

resolverQueryLogConfigAssociation_id :: Lens' ResolverQueryLogConfigAssociation (Maybe Text)

The ID of the query logging association.

resolverQueryLogConfigAssociation_errorMessage :: Lens' ResolverQueryLogConfigAssociation (Maybe Text)

Contains additional information about the error. If the value or Error is null, the value of ErrorMessage also is null.