Network.AWS.Route53Resolver.UpdateResolverRule

Description

Updates settings for a specified Resolver rule. ResolverRuleId is required, and all other parameters are optional. If you don't specify a parameter, it retains its current value.

Creating a Request

data UpdateResolverRule

See: newUpdateResolverRule smart constructor.

Constructors

UpdateResolverRule'

Fields

resolverRuleId :: Text

The ID of the Resolver rule that you want to update.

config :: ResolverRuleConfig

The new settings for the Resolver rule.

newUpdateResolverRule

Arguments

:: Text
$sel:resolverRuleId:UpdateResolverRule'

-> ResolverRuleConfig
$sel:config:UpdateResolverRule'

-> UpdateResolverRule

Create a value of UpdateResolverRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolverRuleId:UpdateResolverRule', updateResolverRule_resolverRuleId - The ID of the Resolver rule that you want to update.

$sel:config:UpdateResolverRule', updateResolverRule_config - The new settings for the Resolver rule.

Request Lenses

updateResolverRule_resolverRuleId :: Lens' UpdateResolverRule Text

The ID of the Resolver rule that you want to update.

updateResolverRule_config :: Lens' UpdateResolverRule ResolverRuleConfig

The new settings for the Resolver rule.

Destructuring the Response

data UpdateResolverRuleResponse

See: newUpdateResolverRuleResponse smart constructor.

Constructors

UpdateResolverRuleResponse'

Fields

resolverRule :: Maybe ResolverRule

The response to an UpdateResolverRule request.

httpStatus :: Int

The response's http status code.

newUpdateResolverRuleResponse

Arguments

:: Int
$sel:httpStatus:UpdateResolverRuleResponse'

-> UpdateResolverRuleResponse

Create a value of UpdateResolverRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolverRule:UpdateResolverRuleResponse', updateResolverRuleResponse_resolverRule - The response to an UpdateResolverRule request.

$sel:httpStatus:UpdateResolverRuleResponse', updateResolverRuleResponse_httpStatus - The response's http status code.

Response Lenses

updateResolverRuleResponse_resolverRule :: Lens' UpdateResolverRuleResponse (Maybe ResolverRule)

The response to an UpdateResolverRule request.

updateResolverRuleResponse_httpStatus :: Lens' UpdateResolverRuleResponse Int

The response's http status code.