Network.AWS.S3

Description

Amazon Simple Storage Service (Amazon S3) is a web service that enables you to store data in the cloud.

You can then download the data or use the data with other AWS services, such as Amazon Elastic Compute Cloud (Amazon EC2).

Operations on the Service

GET Service

Operations on Buckets

GET Bucket

data GetBucket

This implementation of the GET operation returns some or all (up to 1000) of the objects in a bucket. You can use the request parameters as selection criteria to return a subset of the objects in a bucket.

To use this implementation of the operation, you must have READ access to the bucket.

http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketGET.html

Constructors

GetBucket

Fields

gbBucket :: !Text

Target bucket.

gbDelimiter :: !Delimiter

A delimiter is a character you use to group keys. All keys that contain the same string between the prefix, if specified, and the first occurrence of the delimiter after the prefix are grouped under a single result element, CommonPrefixes.

gbPrefix :: Maybe Text

Limits the response to keys that begin with the specified prefix. You can use prefixes to separate a bucket into different groupings of keys. (You can think of using prefix to make groups in the same way you'd use a folder in a file system.)

gbMaxKeys :: !Int

Sets the maximum number of keys returned in the response body. The response might contain fewer keys but will never contain more. Use pagination to access additional matching keys.

gbMarker :: Maybe Text

Specifies the key to start with when listing objects in a bucket. Amazon S3 lists objects in alphabetical order.

data GetBucketResponse

Constructors

GetBucketResponse

Fields

gbrName :: !Text

gbrPrefix :: Maybe Text

gbrMarker :: Maybe Text

gbrMaxKeys :: !Int

gbrIsTruncated :: !Bool

gbrContents :: [Contents]

Operations on Objects

DELETE Object

data DeleteObject

Removes the null version (if there is one) of an object and inserts a delete marker, which becomes the latest version of the object.

If there isn't a null version, Amazon S3 does not remove any objects.

http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectDELETE.html

Constructors

DeleteObject

Fields

doBucket :: !Text

doKey :: !Text

doHeaders :: [Header]

type DeleteObjectResponse = S3Response

GET Object

data GetObject

Retrieves an object from Amazon S3.

You must have READ access to the object.

If you grant READ access to the anonymous user, you can return the object without using an authorization header.

http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectGET.html

Constructors

GetObject

Fields

goBucket :: !Text

goKey :: !Text

goHeaders :: [Header]

type GetObjectResponse = S3Response

HEAD Object

data HeadObject

Retrieves metadata from an object without returning the object itself.

You must have READ access to the object.

By default, the HEAD operation retrieves metadata from the latest version of an object. If the latest version is a delete marker, Amazon S3 behaves as if the object was deleted. To retrieve metadata from a different version, use the versionId subresource.

http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectHEAD.html

Constructors

HeadObject

Fields

hoBucket :: !Text

hoKey :: !Text

hoHeaders :: [Header]

type HeadObjectResponse = S3Response

PUT Object

data PutObject

Add an object to a bucket.

You must have WRITE permissions on a bucket to add an object to it.

Amazon S3 never adds partial objects; if you receive a success response, Amazon S3 added the entire object to the bucket.

Amazon S3 is a distributed system. If it receives multiple write requests for the same object simultaneously, it overwrites all but the last object written.

Amazon S3 does not provide object locking; if you need this, make sure to build it into your application layer or use versioning instead.

If you enable versioning for a bucket, Amazon S3 automatically generates a unique version ID for the object being stored. Amazon S3 returns this ID in the response using the x-amz-version-id response header.

If versioning is suspended, Amazon S3 always uses null as the version ID for the object stored.

http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectPUT.html

Constructors

PutObject

Fields

poBucket :: !Text

poKey :: !Text

poHeaders :: [Header]

poBody :: !RequestBody

type PutObjectResponse = S3Response

Data Types

s3 :: ByteString -> Service

Currently supported version of the S3 service.

s3NS :: ByteString

XML namespace to annotate S3 elements with. FIXME: Some S3 namespaces have the doc.* prefix?

s3Elem :: ByteString -> NName ByteString

Helper to define S3 namespaced XML elements.

data S3ErrorResponse

Constructors

S3ErrorResponse

Fields

sssError :: !Text

data Bucket

Constructors

Bucket

Fields

bName :: !Text

Bucket's name.

bCreationDate :: !UTCTime

Date the bucket was created.

newtype Delimiter

Constructors

Delimiter

Fields

unDelimiter :: Char

data Contents

Constructors

Contents

Fields

bcKey :: !Text

bcLastModified :: !UTCTime

bcETag :: !ETag

bcSize :: !Integer

bcStorageClass :: !StorageClass

data DMObject

Constructors

DMObject

Fields

oKey :: !Text

oVersion :: Maybe Text

data DMObjects

Constructors

DMObjects

Fields

oQuiet :: !Bool

oObjects :: [DMObject]

data DeletedObject

Constructors

DeletedObject

Fields

dobjKey :: !Text

dobjVersionId :: Maybe Text

dobjDeleteMarker :: Maybe Bool

dobjDeleteMarkerVersionId :: Maybe Text

data DeleteError

Constructors

DeleteError

Fields

deKey :: !Text

deVersionId :: Maybe Text

deCode :: !Text

deMessage :: !Text

data Owner

Constructors

Owner

Fields

oID :: !Text

Bucket owner's user ID.

oDisplayName :: !Text

Bucket owner's display name.

data Grantee

Constructors

Grantee

Fields

gID :: !Text

gDisplayName :: !Text

data Grant

Constructors

Grant

Fields

gGrantee :: !Grantee

gPermission :: !Text

data AccessControlList

Constructors

AccessControlList

Fields

aclGrant :: [Grant]

data AccessControlPolicy

Constructors

AccessControlPolicy

Fields

acpOwner :: !Owner

acpAccessControlList :: !AccessControlList

data RestoreRequest

Constructors

RestoreRequest

Fields

rrDays :: !Int

data Part

Constructors

Part

Fields

pPartNumber :: !Text

pETag :: !ETag

data AES256

Constructors

AES256

data CannedACL

Constructors

Private

PublicRead

PublicReadWrite

AuthenticatedRead

BucketOwnerRead

BucketOwnerFullControl

LogDeliveryWrite

data StorageClass

Constructors

Standard

ReducedRedundancy

Glacier

data Directive

Constructors

Copy

Replace

data Source

Constructors

Source

Fields

srcBucket :: !Text

srcKey :: !Text

newtype ETag

Constructors

ETag

Fields

unETag :: Text

Common

module Network.AWS