Network.AWS.S3.AbortMultipartUpload

Description

This action aborts a multipart upload. After a multipart upload is aborted, no additional parts can be uploaded using that upload ID. The storage consumed by any previously uploaded parts will be freed. However, if any part uploads are currently in progress, those part uploads might or might not succeed. As a result, it might be necessary to abort a given multipart upload multiple times in order to completely free all storage consumed by all parts.

To verify that all parts have been removed, so you don't get charged for the part storage, you should call the ListParts action and ensure that the parts list is empty.

For information about permissions required to use the multipart upload, see Multipart Upload and Permissions.

The following operations are related to AbortMultipartUpload:

CreateMultipartUpload

UploadPart

CompleteMultipartUpload

ListParts

ListMultipartUploads

Creating a Request

data AbortMultipartUpload

See: newAbortMultipartUpload smart constructor.

Constructors

AbortMultipartUpload'

Fields

requestPayer :: Maybe RequestPayer

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The bucket name to which the upload was taking place.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

key :: ObjectKey

Key of the object for which the multipart upload was initiated.

uploadId :: Text

Upload ID that identifies the multipart upload.

newAbortMultipartUpload

Arguments

:: BucketName
$sel:bucket:AbortMultipartUpload'

-> ObjectKey
$sel:key:AbortMultipartUpload'

-> Text
$sel:uploadId:AbortMultipartUpload'

-> AbortMultipartUpload

Create a value of AbortMultipartUpload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestPayer:AbortMultipartUpload', abortMultipartUpload_requestPayer - Undocumented member.

$sel:expectedBucketOwner:AbortMultipartUpload', abortMultipartUpload_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:AbortMultipartUpload', abortMultipartUpload_bucket - The bucket name to which the upload was taking place.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

$sel:key:AbortMultipartUpload', abortMultipartUpload_key - Key of the object for which the multipart upload was initiated.

$sel:uploadId:AbortMultipartUpload', abortMultipartUpload_uploadId - Upload ID that identifies the multipart upload.

Request Lenses

abortMultipartUpload_requestPayer :: Lens' AbortMultipartUpload (Maybe RequestPayer)

Undocumented member.

abortMultipartUpload_expectedBucketOwner :: Lens' AbortMultipartUpload (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

abortMultipartUpload_bucket :: Lens' AbortMultipartUpload BucketName

The bucket name to which the upload was taking place.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

abortMultipartUpload_key :: Lens' AbortMultipartUpload ObjectKey

Key of the object for which the multipart upload was initiated.

abortMultipartUpload_uploadId :: Lens' AbortMultipartUpload Text

Upload ID that identifies the multipart upload.

Destructuring the Response

data AbortMultipartUploadResponse

See: newAbortMultipartUploadResponse smart constructor.

Constructors

AbortMultipartUploadResponse'

Fields

requestCharged :: Maybe RequestCharged

httpStatus :: Int

The response's http status code.

newAbortMultipartUploadResponse

Arguments

:: Int
$sel:httpStatus:AbortMultipartUploadResponse'

-> AbortMultipartUploadResponse

Create a value of AbortMultipartUploadResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestCharged:AbortMultipartUploadResponse', abortMultipartUploadResponse_requestCharged - Undocumented member.

$sel:httpStatus:AbortMultipartUploadResponse', abortMultipartUploadResponse_httpStatus - The response's http status code.

Response Lenses

abortMultipartUploadResponse_requestCharged :: Lens' AbortMultipartUploadResponse (Maybe RequestCharged)

Undocumented member.

abortMultipartUploadResponse_httpStatus :: Lens' AbortMultipartUploadResponse Int

The response's http status code.