Network.AWS.S3.CopyObject

Description

Creates a copy of an object that is already stored in Amazon S3.

You can store individual objects of up to 5 TB in Amazon S3. You create a copy of your object up to 5 GB in size in a single atomic action using this API. However, to copy an object greater than 5 GB, you must use the multipart upload Upload Part - Copy API. For more information, see Copy Object Using the REST Multipart Upload API.

All copy requests must be authenticated. Additionally, you must have read access to the source object and write access to the destination bucket. For more information, see REST Authentication. Both the Region that you want to copy the object from and the Region that you want to copy the object to must be enabled for your account.

A copy request might return an error when Amazon S3 receives the copy request or while Amazon S3 is copying the files. If the error occurs before the copy action starts, you receive a standard Amazon S3 error. If the error occurs during the copy operation, the error response is embedded in the 200 OK response. This means that a 200 OK response can contain either a success or an error. Design your application to parse the contents of the response and handle it appropriately.

If the copy is successful, you receive a response with information about the copied object.

If the request is an HTTP 1.1 request, the response is chunk encoded. If it were not, it would not contain the content-length, and you would need to read the entire body.

The copy request charge is based on the storage class and Region that you specify for the destination object. For pricing information, see Amazon S3 pricing.

Amazon S3 transfer acceleration does not support cross-Region copies. If you request a cross-Region copy using a transfer acceleration endpoint, you get a 400 Bad Request error. For more information, see Transfer Acceleration.

Metadata

When copying an object, you can preserve all metadata (default) or specify new metadata. However, the ACL is not preserved and is set to private for the user making the request. To override the default ACL setting, specify a new ACL when generating a copy request. For more information, see Using ACLs.

To specify whether you want the object metadata copied from the source object or replaced with metadata provided in the request, you can optionally add the x-amz-metadata-directive header. When you grant permissions, you can use the s3:x-amz-metadata-directive condition key to enforce certain metadata behavior when objects are uploaded. For more information, see Specifying Conditions in a Policy in the Amazon S3 User Guide. For a complete list of Amazon S3-specific condition keys, see Actions, Resources, and Condition Keys for Amazon S3.

x-amz-copy-source-if Headers

To only copy an object under certain conditions, such as whether the Etag matches or whether the object was modified before or after a specified date, use the following request parameters:

x-amz-copy-source-if-match

x-amz-copy-source-if-none-match

x-amz-copy-source-if-unmodified-since

x-amz-copy-source-if-modified-since

If both the x-amz-copy-source-if-match and x-amz-copy-source-if-unmodified-since headers are present in the request and evaluate as follows, Amazon S3 returns 200 OK and copies the data:

x-amz-copy-source-if-match condition evaluates to true

x-amz-copy-source-if-unmodified-since condition evaluates to false

If both the x-amz-copy-source-if-none-match and x-amz-copy-source-if-modified-since headers are present in the request and evaluate as follows, Amazon S3 returns the 412 Precondition Failed response code:

x-amz-copy-source-if-none-match condition evaluates to false

x-amz-copy-source-if-modified-since condition evaluates to true

All headers with the x-amz- prefix, including x-amz-copy-source, must be signed.

Server-side encryption

When you perform a CopyObject operation, you can optionally use the appropriate encryption-related headers to encrypt the object using server-side encryption with Amazon Web Services managed encryption keys (SSE-S3 or SSE-KMS) or a customer-provided encryption key. With server-side encryption, Amazon S3 encrypts your data as it writes it to disks in its data centers and decrypts the data when you access it. For more information about server-side encryption, see Using Server-Side Encryption.

If a target object uses SSE-KMS, you can enable an S3 Bucket Key for the object. For more information, see Amazon S3 Bucket Keys in the Amazon S3 User Guide.

Access Control List (ACL)-Specific Request Headers

When copying an object, you can optionally use headers to grant ACL-based permissions. By default, all objects are private. Only the owner has full access control. When adding a new object, you can grant permissions to individual Amazon Web Services accounts or to predefined groups defined by Amazon S3. These permissions are then added to the ACL on the object. For more information, see Access Control List (ACL) Overview and Managing ACLs Using the REST API.

Storage Class Options

You can use the CopyObject action to change the storage class of an object that is already stored in Amazon S3 using the StorageClass parameter. For more information, see Storage Classes in the Amazon S3 User Guide.

Versioning

By default, x-amz-copy-source identifies the current version of an object to copy. If the current version is a delete marker, Amazon S3 behaves as if the object was deleted. To copy a different version, use the versionId subresource.

If you enable versioning on the target bucket, Amazon S3 generates a unique version ID for the object being copied. This version ID is different from the version ID of the source object. Amazon S3 returns the version ID of the copied object in the x-amz-version-id response header in the response.

If you do not enable versioning or suspend it on the target bucket, the version ID that Amazon S3 generates is always null.

If the source object's storage class is GLACIER, you must restore a copy of this object before you can use it as a source object for the copy operation. For more information, see RestoreObject.

The following operations are related to CopyObject:

PutObject

GetObject

For more information, see Copying Objects.

Creating a Request

data CopyObject

See: newCopyObject smart constructor.

Constructors

CopyObject'

Fields

copySourceIfModifiedSince :: Maybe ISO8601

Copies the object if it has been modified since the specified time.

copySourceIfUnmodifiedSince :: Maybe ISO8601

Copies the object if it hasn't been modified since the specified time.

copySourceSSECustomerKeyMD5 :: Maybe Text

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

taggingDirective :: Maybe TaggingDirective

Specifies whether the object tag-set are copied from the source object or replaced with tag-set provided in the request.

metadataDirective :: Maybe MetadataDirective

Specifies whether the metadata is copied from the source object or replaced with metadata provided in the request.

objectLockMode :: Maybe ObjectLockMode

The Object Lock mode that you want to apply to the copied object.

expires :: Maybe ISO8601

The date and time at which the object is no longer cacheable.

grantReadACP :: Maybe Text

Allows grantee to read the object ACL.

This action is not supported by Amazon S3 on Outposts.

copySourceIfNoneMatch :: Maybe Text

Copies the object if its entity tag (ETag) is different than the specified ETag.

sSECustomerAlgorithm :: Maybe Text

Specifies the algorithm to use to when encrypting the object (for example, AES256).

sSECustomerKey :: Maybe (Sensitive Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

requestPayer :: Maybe RequestPayer

grantWriteACP :: Maybe Text

Allows grantee to write the ACL for the applicable object.

This action is not supported by Amazon S3 on Outposts.

copySourceIfMatch :: Maybe Text

Copies the object if its entity tag (ETag) matches the specified tag.

bucketKeyEnabled :: Maybe Bool

Specifies whether Amazon S3 should use an S3 Bucket Key for object encryption with server-side encryption using AWS KMS (SSE-KMS). Setting this header to true causes Amazon S3 to use an S3 Bucket Key for object encryption with SSE-KMS.

Specifying this header with a COPY action doesn’t affect bucket-level settings for S3 Bucket Key.

websiteRedirectLocation :: Maybe Text

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

grantRead :: Maybe Text

Allows grantee to read the object data and its metadata.

This action is not supported by Amazon S3 on Outposts.

expectedSourceBucketOwner :: Maybe Text

The account ID of the expected source bucket owner. If the source bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

storageClass :: Maybe StorageClass

By default, Amazon S3 uses the STANDARD Storage Class to store newly created objects. The STANDARD storage class provides high durability and high availability. Depending on performance needs, you can specify a different Storage Class. Amazon S3 on Outposts only uses the OUTPOSTS Storage Class. For more information, see Storage Classes in the Amazon S3 User Guide.

sSECustomerKeyMD5 :: Maybe Text

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

sSEKMSKeyId :: Maybe (Sensitive Text)

Specifies the Amazon Web Services KMS key ID to use for object encryption. All GET and PUT requests for an object protected by Amazon Web Services KMS will fail if not made via SSL or using SigV4. For information about configuring using any of the officially supported Amazon Web Services SDKs and Amazon Web Services CLI, see Specifying the Signature Version in Request Authentication in the Amazon S3 User Guide.

grantFullControl :: Maybe Text

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

This action is not supported by Amazon S3 on Outposts.

contentEncoding :: Maybe Text

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

tagging :: Maybe Text

The tag-set for the object destination object this value must be used in conjunction with the TaggingDirective. The tag-set must be encoded as URL Query parameters.

objectLockRetainUntilDate :: Maybe ISO8601

The date and time when you want the copied object's Object Lock to expire.

metadata :: HashMap Text Text

A map of metadata to store with the object in S3.

sSEKMSEncryptionContext :: Maybe (Sensitive Text)

Specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

cacheControl :: Maybe Text

Specifies caching behavior along the request/reply chain.

contentLanguage :: Maybe Text

The language the content is in.

copySourceSSECustomerKey :: Maybe (Sensitive Text)

Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

objectLockLegalHoldStatus :: Maybe ObjectLockLegalHoldStatus

Specifies whether you want to apply a Legal Hold to the copied object.

copySourceSSECustomerAlgorithm :: Maybe Text

Specifies the algorithm to use when decrypting the source object (for example, AES256).

acl :: Maybe ObjectCannedACL

The canned ACL to apply to the object.

This action is not supported by Amazon S3 on Outposts.

contentDisposition :: Maybe Text

Specifies presentational information for the object.

expectedBucketOwner :: Maybe Text

The account ID of the expected destination bucket owner. If the destination bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

serverSideEncryption :: Maybe ServerSideEncryption

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

contentType :: Maybe Text

A standard MIME type describing the format of the object data.

bucket :: BucketName

The name of the destination bucket.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

copySource :: Text

Specifies the source object for the copy operation. You specify the value in one of two formats, depending on whether you want to access the source object through an access point:

For objects not accessed through an access point, specify the name of the source bucket and the key of the source object, separated by a slash (/). For example, to copy the object reports/january.pdf from the bucket awsexamplebucket, use awsexamplebucket/reports/january.pdf. The value must be URL encoded.

For objects accessed through access points, specify the Amazon Resource Name (ARN) of the object as accessed through the access point, in the format arn:aws:s3:<Region>:<account-id>:accesspoint/<access-point-name>/object/<key>. For example, to copy the object reports/january.pdf through access point my-access-point owned by account 123456789012 in Region us-west-2, use the URL encoding of arn:aws:s3:us-west-2:123456789012:accesspoint/my-access-point/object/reports/january.pdf. The value must be URL encoded.

Amazon S3 supports copy operations using access points only when the source and destination buckets are in the same Amazon Web Services Region.

Alternatively, for objects accessed through Amazon S3 on Outposts, specify the ARN of the object as accessed in the format arn:aws:s3-outposts:<Region>:<account-id>:outpost/<outpost-id>/object/<key>. For example, to copy the object reports/january.pdf through outpost my-outpost owned by account 123456789012 in Region us-west-2, use the URL encoding of arn:aws:s3-outposts:us-west-2:123456789012:outpost/my-outpost/object/reports/january.pdf. The value must be URL encoded.

To copy a specific version of an object, append ?versionId=<version-id> to the value (for example, awsexamplebucket/reports/january.pdf?versionId=QUpfdndhfd8438MNFDN93jdnJFkdmqnh893). If you don't specify a version ID, Amazon S3 copies the latest version of the source object.

key :: ObjectKey

The key of the destination object.

newCopyObject

Arguments

:: BucketName
$sel:bucket:CopyObject'

-> Text
$sel:copySource:CopyObject'

-> ObjectKey
$sel:key:CopyObject'

-> CopyObject

Create a value of CopyObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:copySourceIfModifiedSince:CopyObject', copyObject_copySourceIfModifiedSince - Copies the object if it has been modified since the specified time.

$sel:copySourceIfUnmodifiedSince:CopyObject', copyObject_copySourceIfUnmodifiedSince - Copies the object if it hasn't been modified since the specified time.

$sel:copySourceSSECustomerKeyMD5:CopyObject', copyObject_copySourceSSECustomerKeyMD5 - Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

$sel:taggingDirective:CopyObject', copyObject_taggingDirective - Specifies whether the object tag-set are copied from the source object or replaced with tag-set provided in the request.

$sel:metadataDirective:CopyObject', copyObject_metadataDirective - Specifies whether the metadata is copied from the source object or replaced with metadata provided in the request.

$sel:objectLockMode:CopyObject', copyObject_objectLockMode - The Object Lock mode that you want to apply to the copied object.

$sel:expires:CopyObject', copyObject_expires - The date and time at which the object is no longer cacheable.

$sel:grantReadACP:CopyObject', copyObject_grantReadACP - Allows grantee to read the object ACL.

This action is not supported by Amazon S3 on Outposts.

$sel:copySourceIfNoneMatch:CopyObject', copyObject_copySourceIfNoneMatch - Copies the object if its entity tag (ETag) is different than the specified ETag.

$sel:sSECustomerAlgorithm:CopyObject', copyObject_sSECustomerAlgorithm - Specifies the algorithm to use to when encrypting the object (for example, AES256).

$sel:sSECustomerKey:CopyObject', copyObject_sSECustomerKey - Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

$sel:requestPayer:CopyObject', copyObject_requestPayer - Undocumented member.

$sel:grantWriteACP:CopyObject', copyObject_grantWriteACP - Allows grantee to write the ACL for the applicable object.

This action is not supported by Amazon S3 on Outposts.

$sel:copySourceIfMatch:CopyObject', copyObject_copySourceIfMatch - Copies the object if its entity tag (ETag) matches the specified tag.

$sel:bucketKeyEnabled:CopyObject', copyObject_bucketKeyEnabled - Specifies whether Amazon S3 should use an S3 Bucket Key for object encryption with server-side encryption using AWS KMS (SSE-KMS). Setting this header to true causes Amazon S3 to use an S3 Bucket Key for object encryption with SSE-KMS.

Specifying this header with a COPY action doesn’t affect bucket-level settings for S3 Bucket Key.

$sel:websiteRedirectLocation:CopyObject', copyObject_websiteRedirectLocation - If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

$sel:grantRead:CopyObject', copyObject_grantRead - Allows grantee to read the object data and its metadata.

This action is not supported by Amazon S3 on Outposts.

$sel:expectedSourceBucketOwner:CopyObject', copyObject_expectedSourceBucketOwner - The account ID of the expected source bucket owner. If the source bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:storageClass:CopyObject', copyObject_storageClass - By default, Amazon S3 uses the STANDARD Storage Class to store newly created objects. The STANDARD storage class provides high durability and high availability. Depending on performance needs, you can specify a different Storage Class. Amazon S3 on Outposts only uses the OUTPOSTS Storage Class. For more information, see Storage Classes in the Amazon S3 User Guide.

$sel:sSECustomerKeyMD5:CopyObject', copyObject_sSECustomerKeyMD5 - Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

$sel:sSEKMSKeyId:CopyObject', copyObject_sSEKMSKeyId - Specifies the Amazon Web Services KMS key ID to use for object encryption. All GET and PUT requests for an object protected by Amazon Web Services KMS will fail if not made via SSL or using SigV4. For information about configuring using any of the officially supported Amazon Web Services SDKs and Amazon Web Services CLI, see Specifying the Signature Version in Request Authentication in the Amazon S3 User Guide.

$sel:grantFullControl:CopyObject', copyObject_grantFullControl - Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

This action is not supported by Amazon S3 on Outposts.

$sel:contentEncoding:CopyObject', copyObject_contentEncoding - Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

$sel:tagging:CopyObject', copyObject_tagging - The tag-set for the object destination object this value must be used in conjunction with the TaggingDirective. The tag-set must be encoded as URL Query parameters.

$sel:objectLockRetainUntilDate:CopyObject', copyObject_objectLockRetainUntilDate - The date and time when you want the copied object's Object Lock to expire.

$sel:metadata:CopyObject', copyObject_metadata - A map of metadata to store with the object in S3.

$sel:sSEKMSEncryptionContext:CopyObject', copyObject_sSEKMSEncryptionContext - Specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

$sel:cacheControl:CopyObject', copyObject_cacheControl - Specifies caching behavior along the request/reply chain.

$sel:contentLanguage:CopyObject', copyObject_contentLanguage - The language the content is in.

$sel:copySourceSSECustomerKey:CopyObject', copyObject_copySourceSSECustomerKey - Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

$sel:objectLockLegalHoldStatus:CopyObject', copyObject_objectLockLegalHoldStatus - Specifies whether you want to apply a Legal Hold to the copied object.

$sel:copySourceSSECustomerAlgorithm:CopyObject', copyObject_copySourceSSECustomerAlgorithm - Specifies the algorithm to use when decrypting the source object (for example, AES256).

$sel:acl:CopyObject', copyObject_acl - The canned ACL to apply to the object.

This action is not supported by Amazon S3 on Outposts.

$sel:contentDisposition:CopyObject', copyObject_contentDisposition - Specifies presentational information for the object.

$sel:expectedBucketOwner:CopyObject', copyObject_expectedBucketOwner - The account ID of the expected destination bucket owner. If the destination bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:serverSideEncryption:CopyObject', copyObject_serverSideEncryption - The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

$sel:contentType:CopyObject', copyObject_contentType - A standard MIME type describing the format of the object data.

$sel:bucket:CopyObject', copyObject_bucket - The name of the destination bucket.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

$sel:copySource:CopyObject', copyObject_copySource - Specifies the source object for the copy operation. You specify the value in one of two formats, depending on whether you want to access the source object through an access point:

For objects not accessed through an access point, specify the name of the source bucket and the key of the source object, separated by a slash (/). For example, to copy the object reports/january.pdf from the bucket awsexamplebucket, use awsexamplebucket/reports/january.pdf. The value must be URL encoded.

For objects accessed through access points, specify the Amazon Resource Name (ARN) of the object as accessed through the access point, in the format arn:aws:s3:<Region>:<account-id>:accesspoint/<access-point-name>/object/<key>. For example, to copy the object reports/january.pdf through access point my-access-point owned by account 123456789012 in Region us-west-2, use the URL encoding of arn:aws:s3:us-west-2:123456789012:accesspoint/my-access-point/object/reports/january.pdf. The value must be URL encoded.

Amazon S3 supports copy operations using access points only when the source and destination buckets are in the same Amazon Web Services Region.

Alternatively, for objects accessed through Amazon S3 on Outposts, specify the ARN of the object as accessed in the format arn:aws:s3-outposts:<Region>:<account-id>:outpost/<outpost-id>/object/<key>. For example, to copy the object reports/january.pdf through outpost my-outpost owned by account 123456789012 in Region us-west-2, use the URL encoding of arn:aws:s3-outposts:us-west-2:123456789012:outpost/my-outpost/object/reports/january.pdf. The value must be URL encoded.

To copy a specific version of an object, append ?versionId=<version-id> to the value (for example, awsexamplebucket/reports/january.pdf?versionId=QUpfdndhfd8438MNFDN93jdnJFkdmqnh893). If you don't specify a version ID, Amazon S3 copies the latest version of the source object.

$sel:key:CopyObject', copyObject_key - The key of the destination object.

Request Lenses

copyObject_copySourceIfModifiedSince :: Lens' CopyObject (Maybe UTCTime)

Copies the object if it has been modified since the specified time.

copyObject_copySourceIfUnmodifiedSince :: Lens' CopyObject (Maybe UTCTime)

Copies the object if it hasn't been modified since the specified time.

copyObject_copySourceSSECustomerKeyMD5 :: Lens' CopyObject (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

copyObject_taggingDirective :: Lens' CopyObject (Maybe TaggingDirective)

Specifies whether the object tag-set are copied from the source object or replaced with tag-set provided in the request.

copyObject_metadataDirective :: Lens' CopyObject (Maybe MetadataDirective)

Specifies whether the metadata is copied from the source object or replaced with metadata provided in the request.

copyObject_objectLockMode :: Lens' CopyObject (Maybe ObjectLockMode)

The Object Lock mode that you want to apply to the copied object.

copyObject_expires :: Lens' CopyObject (Maybe UTCTime)

The date and time at which the object is no longer cacheable.

copyObject_grantReadACP :: Lens' CopyObject (Maybe Text)

Allows grantee to read the object ACL.

This action is not supported by Amazon S3 on Outposts.

copyObject_copySourceIfNoneMatch :: Lens' CopyObject (Maybe Text)

Copies the object if its entity tag (ETag) is different than the specified ETag.

copyObject_sSECustomerAlgorithm :: Lens' CopyObject (Maybe Text)

Specifies the algorithm to use to when encrypting the object (for example, AES256).

copyObject_sSECustomerKey :: Lens' CopyObject (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

copyObject_requestPayer :: Lens' CopyObject (Maybe RequestPayer)

Undocumented member.

copyObject_grantWriteACP :: Lens' CopyObject (Maybe Text)

Allows grantee to write the ACL for the applicable object.

This action is not supported by Amazon S3 on Outposts.

copyObject_copySourceIfMatch :: Lens' CopyObject (Maybe Text)

Copies the object if its entity tag (ETag) matches the specified tag.

copyObject_bucketKeyEnabled :: Lens' CopyObject (Maybe Bool)

Specifies whether Amazon S3 should use an S3 Bucket Key for object encryption with server-side encryption using AWS KMS (SSE-KMS). Setting this header to true causes Amazon S3 to use an S3 Bucket Key for object encryption with SSE-KMS.

Specifying this header with a COPY action doesn’t affect bucket-level settings for S3 Bucket Key.

copyObject_websiteRedirectLocation :: Lens' CopyObject (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

copyObject_grantRead :: Lens' CopyObject (Maybe Text)

Allows grantee to read the object data and its metadata.

This action is not supported by Amazon S3 on Outposts.

copyObject_expectedSourceBucketOwner :: Lens' CopyObject (Maybe Text)

The account ID of the expected source bucket owner. If the source bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

copyObject_storageClass :: Lens' CopyObject (Maybe StorageClass)

By default, Amazon S3 uses the STANDARD Storage Class to store newly created objects. The STANDARD storage class provides high durability and high availability. Depending on performance needs, you can specify a different Storage Class. Amazon S3 on Outposts only uses the OUTPOSTS Storage Class. For more information, see Storage Classes in the Amazon S3 User Guide.

copyObject_sSECustomerKeyMD5 :: Lens' CopyObject (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

copyObject_sSEKMSKeyId :: Lens' CopyObject (Maybe Text)

Specifies the Amazon Web Services KMS key ID to use for object encryption. All GET and PUT requests for an object protected by Amazon Web Services KMS will fail if not made via SSL or using SigV4. For information about configuring using any of the officially supported Amazon Web Services SDKs and Amazon Web Services CLI, see Specifying the Signature Version in Request Authentication in the Amazon S3 User Guide.

copyObject_grantFullControl :: Lens' CopyObject (Maybe Text)

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

This action is not supported by Amazon S3 on Outposts.

copyObject_contentEncoding :: Lens' CopyObject (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

copyObject_tagging :: Lens' CopyObject (Maybe Text)

The tag-set for the object destination object this value must be used in conjunction with the TaggingDirective. The tag-set must be encoded as URL Query parameters.

copyObject_objectLockRetainUntilDate :: Lens' CopyObject (Maybe UTCTime)

The date and time when you want the copied object's Object Lock to expire.

copyObject_metadata :: Lens' CopyObject (HashMap Text Text)

A map of metadata to store with the object in S3.

copyObject_sSEKMSEncryptionContext :: Lens' CopyObject (Maybe Text)

Specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

copyObject_cacheControl :: Lens' CopyObject (Maybe Text)

Specifies caching behavior along the request/reply chain.

copyObject_contentLanguage :: Lens' CopyObject (Maybe Text)

The language the content is in.

copyObject_copySourceSSECustomerKey :: Lens' CopyObject (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

copyObject_objectLockLegalHoldStatus :: Lens' CopyObject (Maybe ObjectLockLegalHoldStatus)

Specifies whether you want to apply a Legal Hold to the copied object.

copyObject_copySourceSSECustomerAlgorithm :: Lens' CopyObject (Maybe Text)

Specifies the algorithm to use when decrypting the source object (for example, AES256).

copyObject_acl :: Lens' CopyObject (Maybe ObjectCannedACL)

The canned ACL to apply to the object.

This action is not supported by Amazon S3 on Outposts.

copyObject_contentDisposition :: Lens' CopyObject (Maybe Text)

Specifies presentational information for the object.

copyObject_expectedBucketOwner :: Lens' CopyObject (Maybe Text)

The account ID of the expected destination bucket owner. If the destination bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

copyObject_serverSideEncryption :: Lens' CopyObject (Maybe ServerSideEncryption)

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

copyObject_contentType :: Lens' CopyObject (Maybe Text)

A standard MIME type describing the format of the object data.

copyObject_bucket :: Lens' CopyObject BucketName

The name of the destination bucket.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

copyObject_copySource :: Lens' CopyObject Text

Specifies the source object for the copy operation. You specify the value in one of two formats, depending on whether you want to access the source object through an access point:

For objects not accessed through an access point, specify the name of the source bucket and the key of the source object, separated by a slash (/). For example, to copy the object reports/january.pdf from the bucket awsexamplebucket, use awsexamplebucket/reports/january.pdf. The value must be URL encoded.

For objects accessed through access points, specify the Amazon Resource Name (ARN) of the object as accessed through the access point, in the format arn:aws:s3:<Region>:<account-id>:accesspoint/<access-point-name>/object/<key>. For example, to copy the object reports/january.pdf through access point my-access-point owned by account 123456789012 in Region us-west-2, use the URL encoding of arn:aws:s3:us-west-2:123456789012:accesspoint/my-access-point/object/reports/january.pdf. The value must be URL encoded.

Amazon S3 supports copy operations using access points only when the source and destination buckets are in the same Amazon Web Services Region.

Alternatively, for objects accessed through Amazon S3 on Outposts, specify the ARN of the object as accessed in the format arn:aws:s3-outposts:<Region>:<account-id>:outpost/<outpost-id>/object/<key>. For example, to copy the object reports/january.pdf through outpost my-outpost owned by account 123456789012 in Region us-west-2, use the URL encoding of arn:aws:s3-outposts:us-west-2:123456789012:outpost/my-outpost/object/reports/january.pdf. The value must be URL encoded.

To copy a specific version of an object, append ?versionId=<version-id> to the value (for example, awsexamplebucket/reports/january.pdf?versionId=QUpfdndhfd8438MNFDN93jdnJFkdmqnh893). If you don't specify a version ID, Amazon S3 copies the latest version of the source object.

copyObject_key :: Lens' CopyObject ObjectKey

The key of the destination object.

Destructuring the Response

data CopyObjectResponse

See: newCopyObjectResponse smart constructor.

Constructors

CopyObjectResponse'

Fields

requestCharged :: Maybe RequestCharged

versionId :: Maybe ObjectVersionId

Version ID of the newly created copy.

expiration :: Maybe Text

If the object expiration is configured, the response includes this header.

sSECustomerAlgorithm :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

bucketKeyEnabled :: Maybe Bool

Indicates whether the copied object uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

copySourceVersionId :: Maybe Text

Version of the copied object in the destination bucket.

sSECustomerKeyMD5 :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

sSEKMSKeyId :: Maybe (Sensitive Text)

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

sSEKMSEncryptionContext :: Maybe (Sensitive Text)

If present, specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

serverSideEncryption :: Maybe ServerSideEncryption

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

copyObjectResult :: Maybe CopyObjectResult

Container for all response elements.

httpStatus :: Int

The response's http status code.

newCopyObjectResponse

Arguments

:: Int
$sel:httpStatus:CopyObjectResponse'

-> CopyObjectResponse

Create a value of CopyObjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestCharged:CopyObjectResponse', copyObjectResponse_requestCharged - Undocumented member.

$sel:versionId:CopyObjectResponse', copyObjectResponse_versionId - Version ID of the newly created copy.

$sel:expiration:CopyObjectResponse', copyObjectResponse_expiration - If the object expiration is configured, the response includes this header.

$sel:sSECustomerAlgorithm:CopyObject', copyObjectResponse_sSECustomerAlgorithm - If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

$sel:bucketKeyEnabled:CopyObject', copyObjectResponse_bucketKeyEnabled - Indicates whether the copied object uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

$sel:copySourceVersionId:CopyObjectResponse', copyObjectResponse_copySourceVersionId - Version of the copied object in the destination bucket.

$sel:sSECustomerKeyMD5:CopyObject', copyObjectResponse_sSECustomerKeyMD5 - If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

$sel:sSEKMSKeyId:CopyObject', copyObjectResponse_sSEKMSKeyId - If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

$sel:sSEKMSEncryptionContext:CopyObject', copyObjectResponse_sSEKMSEncryptionContext - If present, specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

$sel:serverSideEncryption:CopyObject', copyObjectResponse_serverSideEncryption - The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

$sel:copyObjectResult:CopyObjectResponse', copyObjectResponse_copyObjectResult - Container for all response elements.

$sel:httpStatus:CopyObjectResponse', copyObjectResponse_httpStatus - The response's http status code.

Response Lenses

copyObjectResponse_requestCharged :: Lens' CopyObjectResponse (Maybe RequestCharged)

Undocumented member.

copyObjectResponse_versionId :: Lens' CopyObjectResponse (Maybe ObjectVersionId)

Version ID of the newly created copy.

copyObjectResponse_expiration :: Lens' CopyObjectResponse (Maybe Text)

If the object expiration is configured, the response includes this header.

copyObjectResponse_sSECustomerAlgorithm :: Lens' CopyObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

copyObjectResponse_bucketKeyEnabled :: Lens' CopyObjectResponse (Maybe Bool)

Indicates whether the copied object uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

copyObjectResponse_copySourceVersionId :: Lens' CopyObjectResponse (Maybe Text)

Version of the copied object in the destination bucket.

copyObjectResponse_sSECustomerKeyMD5 :: Lens' CopyObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

copyObjectResponse_sSEKMSKeyId :: Lens' CopyObjectResponse (Maybe Text)

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

copyObjectResponse_sSEKMSEncryptionContext :: Lens' CopyObjectResponse (Maybe Text)

If present, specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

copyObjectResponse_serverSideEncryption :: Lens' CopyObjectResponse (Maybe ServerSideEncryption)

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

copyObjectResponse_copyObjectResult :: Lens' CopyObjectResponse (Maybe CopyObjectResult)

Container for all response elements.

copyObjectResponse_httpStatus :: Lens' CopyObjectResponse Int

The response's http status code.