Network.AWS.S3.CreateBucket

Description

Creates a new S3 bucket. To create a bucket, you must register with Amazon S3 and have a valid Amazon Web Services Access Key ID to authenticate requests. Anonymous requests are never allowed to create buckets. By creating the bucket, you become the bucket owner.

Not every string is an acceptable bucket name. For information about bucket naming restrictions, see Bucket naming rules.

If you want to create an Amazon S3 on Outposts bucket, see Create Bucket.

By default, the bucket is created in the US East (N. Virginia) Region. You can optionally specify a Region in the request body. You might choose a Region to optimize latency, minimize costs, or address regulatory requirements. For example, if you reside in Europe, you will probably find it advantageous to create buckets in the Europe (Ireland) Region. For more information, see Accessing a bucket.

If you send your create bucket request to the s3.amazonaws.com endpoint, the request goes to the us-east-1 Region. Accordingly, the signature calculations in Signature Version 4 must use us-east-1 as the Region, even if the location constraint in the request specifies another Region where the bucket is to be created. If you create a bucket in a Region other than US East (N. Virginia), your application must be able to handle 307 redirect. For more information, see Virtual hosting of buckets.

When creating a bucket using this operation, you can optionally specify the accounts or groups that should be granted specific permissions on the bucket. There are two ways to grant the appropriate permissions using the request headers.

Specify a canned ACL using the x-amz-acl request header. Amazon S3 supports a set of predefined ACLs, known as canned ACLs. Each canned ACL has a predefined set of grantees and permissions. For more information, see Canned ACL.

Specify access permissions explicitly using the x-amz-grant-read, x-amz-grant-write, x-amz-grant-read-acp, x-amz-grant-write-acp, and x-amz-grant-full-control headers. These headers map to the set of permissions Amazon S3 supports in an ACL. For more information, see Access control list (ACL) overview.

You specify each grantee as a type=value pair, where the type is one of the following:

id – if the value specified is the canonical user ID of an Amazon Web Services account

uri – if you are granting permissions to a predefined group

emailAddress – if the value specified is the email address of an Amazon Web Services account

Using email addresses to specify a grantee is only supported in the following Amazon Web Services Regions:

US East (N. Virginia)

US West (N. California)

US West (Oregon)

Asia Pacific (Singapore)

Asia Pacific (Sydney)

Asia Pacific (Tokyo)

Europe (Ireland)

South America (São Paulo)

For a list of all the Amazon S3 supported Regions and endpoints, see Regions and Endpoints in the Amazon Web Services General Reference.

For example, the following x-amz-grant-read header grants the Amazon Web Services accounts identified by account IDs permissions to read object data and its metadata:

x-amz-grant-read: id="11112222333", id="444455556666"

You can use either a canned ACL or specify access permissions explicitly. You cannot do both.

Permissions

If your CreateBucket request specifies ACL permissions and the ACL is public-read, public-read-write, authenticated-read, or if you specify access permissions explicitly through any other ACL, both s3:CreateBucket and s3:PutBucketAcl permissions are needed. If the ACL the CreateBucket request is private, only s3:CreateBucket permission is needed.

If ObjectLockEnabledForBucket is set to true in your CreateBucket request, s3:PutBucketObjectLockConfiguration and s3:PutBucketVersioning permissions are required.

The following operations are related to CreateBucket:

PutObject

DeleteBucket

Creating a Request

data CreateBucket

See: newCreateBucket smart constructor.

Constructors

CreateBucket'

Fields

grantReadACP :: Maybe Text

Allows grantee to read the bucket ACL.

objectLockEnabledForBucket :: Maybe Bool

Specifies whether you want S3 Object Lock to be enabled for the new bucket.

grantWriteACP :: Maybe Text

Allows grantee to write the ACL for the applicable bucket.

grantRead :: Maybe Text

Allows grantee to list the objects in the bucket.

grantFullControl :: Maybe Text

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

createBucketConfiguration :: Maybe CreateBucketConfiguration

The configuration information for the bucket.

grantWrite :: Maybe Text

Allows grantee to create new objects in the bucket.

For the bucket and object owners of existing objects, also allows deletions and overwrites of those objects.

acl :: Maybe BucketCannedACL

The canned ACL to apply to the bucket.

bucket :: BucketName

The name of the bucket to create.

newCreateBucket

Arguments

:: BucketName
$sel:bucket:CreateBucket'

-> CreateBucket

Create a value of CreateBucket with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:grantReadACP:CreateBucket', createBucket_grantReadACP - Allows grantee to read the bucket ACL.

$sel:objectLockEnabledForBucket:CreateBucket', createBucket_objectLockEnabledForBucket - Specifies whether you want S3 Object Lock to be enabled for the new bucket.

$sel:grantWriteACP:CreateBucket', createBucket_grantWriteACP - Allows grantee to write the ACL for the applicable bucket.

$sel:grantRead:CreateBucket', createBucket_grantRead - Allows grantee to list the objects in the bucket.

$sel:grantFullControl:CreateBucket', createBucket_grantFullControl - Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

$sel:createBucketConfiguration:CreateBucket', createBucket_createBucketConfiguration - The configuration information for the bucket.

$sel:grantWrite:CreateBucket', createBucket_grantWrite - Allows grantee to create new objects in the bucket.

For the bucket and object owners of existing objects, also allows deletions and overwrites of those objects.

$sel:acl:CreateBucket', createBucket_acl - The canned ACL to apply to the bucket.

$sel:bucket:CreateBucket', createBucket_bucket - The name of the bucket to create.

Request Lenses

createBucket_grantReadACP :: Lens' CreateBucket (Maybe Text)

Allows grantee to read the bucket ACL.

createBucket_objectLockEnabledForBucket :: Lens' CreateBucket (Maybe Bool)

Specifies whether you want S3 Object Lock to be enabled for the new bucket.

createBucket_grantWriteACP :: Lens' CreateBucket (Maybe Text)

Allows grantee to write the ACL for the applicable bucket.

createBucket_grantRead :: Lens' CreateBucket (Maybe Text)

Allows grantee to list the objects in the bucket.

createBucket_grantFullControl :: Lens' CreateBucket (Maybe Text)

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

createBucket_createBucketConfiguration :: Lens' CreateBucket (Maybe CreateBucketConfiguration)

The configuration information for the bucket.

createBucket_grantWrite :: Lens' CreateBucket (Maybe Text)

Allows grantee to create new objects in the bucket.

For the bucket and object owners of existing objects, also allows deletions and overwrites of those objects.

createBucket_acl :: Lens' CreateBucket (Maybe BucketCannedACL)

The canned ACL to apply to the bucket.

createBucket_bucket :: Lens' CreateBucket BucketName

The name of the bucket to create.

Destructuring the Response

data CreateBucketResponse

See: newCreateBucketResponse smart constructor.

Constructors

CreateBucketResponse'

Fields

location :: Maybe Text

Specifies the Region where the bucket will be created. If you are creating a bucket on the US East (N. Virginia) Region (us-east-1), you do not need to specify the location.

httpStatus :: Int

The response's http status code.

newCreateBucketResponse

Arguments

:: Int
$sel:httpStatus:CreateBucketResponse'

-> CreateBucketResponse

Create a value of CreateBucketResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:location:CreateBucketResponse', createBucketResponse_location - Specifies the Region where the bucket will be created. If you are creating a bucket on the US East (N. Virginia) Region (us-east-1), you do not need to specify the location.

$sel:httpStatus:CreateBucketResponse', createBucketResponse_httpStatus - The response's http status code.

Response Lenses

createBucketResponse_location :: Lens' CreateBucketResponse (Maybe Text)

Specifies the Region where the bucket will be created. If you are creating a bucket on the US East (N. Virginia) Region (us-east-1), you do not need to specify the location.

createBucketResponse_httpStatus :: Lens' CreateBucketResponse Int

The response's http status code.