Network.AWS.S3.CreateMultipartUpload

Description

This action initiates a multipart upload and returns an upload ID. This upload ID is used to associate all of the parts in the specific multipart upload. You specify this upload ID in each of your subsequent upload part requests (see UploadPart). You also include this upload ID in the final request to either complete or abort the multipart upload request.

For more information about multipart uploads, see Multipart Upload Overview.

If you have configured a lifecycle rule to abort incomplete multipart uploads, the upload must complete within the number of days specified in the bucket lifecycle configuration. Otherwise, the incomplete multipart upload becomes eligible for an abort action and Amazon S3 aborts the multipart upload. For more information, see Aborting Incomplete Multipart Uploads Using a Bucket Lifecycle Policy.

For information about the permissions required to use the multipart upload API, see Multipart Upload and Permissions.

For request signing, multipart upload is just a series of regular requests. You initiate a multipart upload, send one or more requests to upload parts, and then complete the multipart upload process. You sign each request individually. There is nothing special about signing multipart upload requests. For more information about signing, see Authenticating Requests (Amazon Web Services Signature Version 4).

After you initiate a multipart upload and upload one or more parts, to stop being charged for storing the uploaded parts, you must either complete or abort the multipart upload. Amazon S3 frees up the space used to store the parts and stop charging you for storing them only after you either complete or abort a multipart upload.

You can optionally request server-side encryption. For server-side encryption, Amazon S3 encrypts your data as it writes it to disks in its data centers and decrypts it when you access it. You can provide your own encryption key, or use Amazon Web Services KMS keys or Amazon S3-managed encryption keys. If you choose to provide your own encryption key, the request headers you provide in UploadPart and UploadPartCopy requests must match the headers you used in the request to initiate the upload by using CreateMultipartUpload.

To perform a multipart upload with encryption using an Amazon Web Services KMS key, the requester must have permission to the kms:Decrypt and kms:GenerateDataKey* actions on the key. These permissions are required because Amazon S3 must decrypt and read data from the encrypted file parts before it completes the multipart upload. For more information, see Multipart upload API and permissions in the Amazon S3 User Guide.

If your Identity and Access Management (IAM) user or role is in the same Amazon Web Services account as the KMS key, then you must have these permissions on the key policy. If your IAM user or role belongs to a different account than the key, then you must have the permissions on both the key policy and your IAM user or role.

For more information, see Protecting Data Using Server-Side Encryption.

Access Permissions

When copying an object, you can optionally specify the accounts or groups that should be granted specific permissions on the new object. There are two ways to grant the permissions using the request headers:

Specify a canned ACL with the x-amz-acl request header. For more information, see Canned ACL.

Specify access permissions explicitly with the x-amz-grant-read, x-amz-grant-read-acp, x-amz-grant-write-acp, and x-amz-grant-full-control headers. These parameters map to the set of permissions that Amazon S3 supports in an ACL. For more information, see Access Control List (ACL) Overview.

You can use either a canned ACL or specify access permissions explicitly. You cannot do both.

Server-Side- Encryption-Specific Request Headers

You can optionally tell Amazon S3 to encrypt data at rest using server-side encryption. Server-side encryption is for data encryption at rest. Amazon S3 encrypts your data as it writes it to disks in its data centers and decrypts it when you access it. The option you use depends on whether you want to use Amazon Web Services managed encryption keys or provide your own encryption key.

Use encryption keys managed by Amazon S3 or customer managed key stored in Amazon Web Services Key Management Service (Amazon Web Services KMS) – If you want Amazon Web Services to manage the keys used to encrypt data, specify the following headers in the request.

x-amz-server-side-encryption

x-amz-server-side-encryption-aws-kms-key-id

x-amz-server-side-encryption-context

If you specify x-amz-server-side-encryption:aws:kms, but don't provide x-amz-server-side-encryption-aws-kms-key-id, Amazon S3 uses the Amazon Web Services managed key in Amazon Web Services KMS to protect the data.

All GET and PUT requests for an object protected by Amazon Web Services KMS fail if you don't make them with SSL or by using SigV4.

For more information about server-side encryption with KMS key (SSE-KMS), see Protecting Data Using Server-Side Encryption with KMS keys.

Use customer-provided encryption keys – If you want to manage your own encryption keys, provide all the following headers in the request.

x-amz-server-side-encryption-customer-algorithm

x-amz-server-side-encryption-customer-key

x-amz-server-side-encryption-customer-key-MD5

For more information about server-side encryption with KMS keys (SSE-KMS), see Protecting Data Using Server-Side Encryption with KMS keys.

Access-Control-List (ACL)-Specific Request Headers

You also can use the following access control–related headers with this operation. By default, all objects are private. Only the owner has full access control. When adding a new object, you can grant permissions to individual Amazon Web Services accounts or to predefined groups defined by Amazon S3. These permissions are then added to the access control list (ACL) on the object. For more information, see Using ACLs. With this operation, you can grant access permissions using one of the following two methods:

Specify a canned ACL (x-amz-acl) — Amazon S3 supports a set of predefined ACLs, known as canned ACLs. Each canned ACL has a predefined set of grantees and permissions. For more information, see Canned ACL.

Specify access permissions explicitly — To explicitly grant access permissions to specific Amazon Web Services accounts or groups, use the following headers. Each header maps to specific permissions that Amazon S3 supports in an ACL. For more information, see Access Control List (ACL) Overview. In the header, you specify a list of grantees who get the specific permission. To grant permissions explicitly, use:

x-amz-grant-read

x-amz-grant-write

x-amz-grant-read-acp

x-amz-grant-write-acp

x-amz-grant-full-control

You specify each grantee as a type=value pair, where the type is one of the following:

id – if the value specified is the canonical user ID of an Amazon Web Services account

uri – if you are granting permissions to a predefined group

emailAddress – if the value specified is the email address of an Amazon Web Services account

Using email addresses to specify a grantee is only supported in the following Amazon Web Services Regions:

US East (N. Virginia)

US West (N. California)

US West (Oregon)

Asia Pacific (Singapore)

Asia Pacific (Sydney)

Asia Pacific (Tokyo)

Europe (Ireland)

South America (São Paulo)

For a list of all the Amazon S3 supported Regions and endpoints, see Regions and Endpoints in the Amazon Web Services General Reference.

For example, the following x-amz-grant-read header grants the Amazon Web Services accounts identified by account IDs permissions to read object data and its metadata:

x-amz-grant-read: id="11112222333", id="444455556666"

The following operations are related to CreateMultipartUpload:

UploadPart

CompleteMultipartUpload

AbortMultipartUpload

ListParts

ListMultipartUploads

Creating a Request

data CreateMultipartUpload

See: newCreateMultipartUpload smart constructor.

Constructors

CreateMultipartUpload'

Fields

grantRead :: Maybe Text

Allows grantee to read the object data and its metadata.

This action is not supported by Amazon S3 on Outposts.

websiteRedirectLocation :: Maybe Text

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

contentType :: Maybe Text

A standard MIME type describing the format of the object data.

contentDisposition :: Maybe Text

Specifies presentational information for the object.

contentLanguage :: Maybe Text

The language the content is in.

sSEKMSEncryptionContext :: Maybe (Sensitive Text)

Specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

metadata :: HashMap Text Text

A map of metadata to store with the object in S3.

sSECustomerKeyMD5 :: Maybe Text

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

sSEKMSKeyId :: Maybe (Sensitive Text)

Specifies the ID of the symmetric customer managed key to use for object encryption. All GET and PUT requests for an object protected by Amazon Web Services KMS will fail if not made via SSL or using SigV4. For information about configuring using any of the officially supported Amazon Web Services SDKs and Amazon Web Services CLI, see Specifying the Signature Version in Request Authentication in the Amazon S3 User Guide.

contentEncoding :: Maybe Text

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

storageClass :: Maybe StorageClass

By default, Amazon S3 uses the STANDARD Storage Class to store newly created objects. The STANDARD storage class provides high durability and high availability. Depending on performance needs, you can specify a different Storage Class. Amazon S3 on Outposts only uses the OUTPOSTS Storage Class. For more information, see Storage Classes in the Amazon S3 User Guide.

bucketKeyEnabled :: Maybe Bool

Specifies whether Amazon S3 should use an S3 Bucket Key for object encryption with server-side encryption using AWS KMS (SSE-KMS). Setting this header to true causes Amazon S3 to use an S3 Bucket Key for object encryption with SSE-KMS.

Specifying this header with an object action doesn’t affect bucket-level settings for S3 Bucket Key.

grantWriteACP :: Maybe Text

Allows grantee to write the ACL for the applicable object.

This action is not supported by Amazon S3 on Outposts.

serverSideEncryption :: Maybe ServerSideEncryption

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

acl :: Maybe ObjectCannedACL

The canned ACL to apply to the object.

This action is not supported by Amazon S3 on Outposts.

sSECustomerKey :: Maybe (Sensitive Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

sSECustomerAlgorithm :: Maybe Text

Specifies the algorithm to use to when encrypting the object (for example, AES256).

grantReadACP :: Maybe Text

Allows grantee to read the object ACL.

This action is not supported by Amazon S3 on Outposts.

objectLockLegalHoldStatus :: Maybe ObjectLockLegalHoldStatus

Specifies whether you want to apply a Legal Hold to the uploaded object.

requestPayer :: Maybe RequestPayer

cacheControl :: Maybe Text

Specifies caching behavior along the request/reply chain.

expires :: Maybe ISO8601

The date and time at which the object is no longer cacheable.

objectLockMode :: Maybe ObjectLockMode

Specifies the Object Lock mode that you want to apply to the uploaded object.

objectLockRetainUntilDate :: Maybe ISO8601

Specifies the date and time when you want the Object Lock to expire.

tagging :: Maybe Text

The tag-set for the object. The tag-set must be encoded as URL Query parameters.

grantFullControl :: Maybe Text

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

This action is not supported by Amazon S3 on Outposts.

bucket :: BucketName

The name of the bucket to which to initiate the upload

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

key :: ObjectKey

Object key for which the multipart upload is to be initiated.

newCreateMultipartUpload

Arguments

:: BucketName
$sel:bucket:CreateMultipartUpload'

-> ObjectKey
$sel:key:CreateMultipartUpload'

-> CreateMultipartUpload

Create a value of CreateMultipartUpload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:grantRead:CreateMultipartUpload', createMultipartUpload_grantRead - Allows grantee to read the object data and its metadata.

This action is not supported by Amazon S3 on Outposts.

$sel:websiteRedirectLocation:CreateMultipartUpload', createMultipartUpload_websiteRedirectLocation - If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

$sel:expectedBucketOwner:CreateMultipartUpload', createMultipartUpload_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:contentType:CreateMultipartUpload', createMultipartUpload_contentType - A standard MIME type describing the format of the object data.

$sel:contentDisposition:CreateMultipartUpload', createMultipartUpload_contentDisposition - Specifies presentational information for the object.

$sel:contentLanguage:CreateMultipartUpload', createMultipartUpload_contentLanguage - The language the content is in.

$sel:sSEKMSEncryptionContext:CreateMultipartUpload', createMultipartUpload_sSEKMSEncryptionContext - Specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

$sel:metadata:CreateMultipartUpload', createMultipartUpload_metadata - A map of metadata to store with the object in S3.

$sel:sSECustomerKeyMD5:CreateMultipartUpload', createMultipartUpload_sSECustomerKeyMD5 - Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

$sel:sSEKMSKeyId:CreateMultipartUpload', createMultipartUpload_sSEKMSKeyId - Specifies the ID of the symmetric customer managed key to use for object encryption. All GET and PUT requests for an object protected by Amazon Web Services KMS will fail if not made via SSL or using SigV4. For information about configuring using any of the officially supported Amazon Web Services SDKs and Amazon Web Services CLI, see Specifying the Signature Version in Request Authentication in the Amazon S3 User Guide.

$sel:contentEncoding:CreateMultipartUpload', createMultipartUpload_contentEncoding - Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

$sel:storageClass:CreateMultipartUpload', createMultipartUpload_storageClass - By default, Amazon S3 uses the STANDARD Storage Class to store newly created objects. The STANDARD storage class provides high durability and high availability. Depending on performance needs, you can specify a different Storage Class. Amazon S3 on Outposts only uses the OUTPOSTS Storage Class. For more information, see Storage Classes in the Amazon S3 User Guide.

$sel:bucketKeyEnabled:CreateMultipartUpload', createMultipartUpload_bucketKeyEnabled - Specifies whether Amazon S3 should use an S3 Bucket Key for object encryption with server-side encryption using AWS KMS (SSE-KMS). Setting this header to true causes Amazon S3 to use an S3 Bucket Key for object encryption with SSE-KMS.

Specifying this header with an object action doesn’t affect bucket-level settings for S3 Bucket Key.

$sel:grantWriteACP:CreateMultipartUpload', createMultipartUpload_grantWriteACP - Allows grantee to write the ACL for the applicable object.

This action is not supported by Amazon S3 on Outposts.

$sel:serverSideEncryption:CreateMultipartUpload', createMultipartUpload_serverSideEncryption - The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

$sel:acl:CreateMultipartUpload', createMultipartUpload_acl - The canned ACL to apply to the object.

This action is not supported by Amazon S3 on Outposts.

$sel:sSECustomerKey:CreateMultipartUpload', createMultipartUpload_sSECustomerKey - Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

$sel:sSECustomerAlgorithm:CreateMultipartUpload', createMultipartUpload_sSECustomerAlgorithm - Specifies the algorithm to use to when encrypting the object (for example, AES256).

$sel:grantReadACP:CreateMultipartUpload', createMultipartUpload_grantReadACP - Allows grantee to read the object ACL.

This action is not supported by Amazon S3 on Outposts.

$sel:objectLockLegalHoldStatus:CreateMultipartUpload', createMultipartUpload_objectLockLegalHoldStatus - Specifies whether you want to apply a Legal Hold to the uploaded object.

$sel:requestPayer:CreateMultipartUpload', createMultipartUpload_requestPayer - Undocumented member.

$sel:cacheControl:CreateMultipartUpload', createMultipartUpload_cacheControl - Specifies caching behavior along the request/reply chain.

$sel:expires:CreateMultipartUpload', createMultipartUpload_expires - The date and time at which the object is no longer cacheable.

$sel:objectLockMode:CreateMultipartUpload', createMultipartUpload_objectLockMode - Specifies the Object Lock mode that you want to apply to the uploaded object.

$sel:objectLockRetainUntilDate:CreateMultipartUpload', createMultipartUpload_objectLockRetainUntilDate - Specifies the date and time when you want the Object Lock to expire.

$sel:tagging:CreateMultipartUpload', createMultipartUpload_tagging - The tag-set for the object. The tag-set must be encoded as URL Query parameters.

$sel:grantFullControl:CreateMultipartUpload', createMultipartUpload_grantFullControl - Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

This action is not supported by Amazon S3 on Outposts.

$sel:bucket:CreateMultipartUpload', createMultipartUpload_bucket - The name of the bucket to which to initiate the upload

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

$sel:key:CreateMultipartUpload', createMultipartUpload_key - Object key for which the multipart upload is to be initiated.

Request Lenses

createMultipartUpload_grantRead :: Lens' CreateMultipartUpload (Maybe Text)

Allows grantee to read the object data and its metadata.

This action is not supported by Amazon S3 on Outposts.

createMultipartUpload_websiteRedirectLocation :: Lens' CreateMultipartUpload (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

createMultipartUpload_expectedBucketOwner :: Lens' CreateMultipartUpload (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

createMultipartUpload_contentType :: Lens' CreateMultipartUpload (Maybe Text)

A standard MIME type describing the format of the object data.

createMultipartUpload_contentDisposition :: Lens' CreateMultipartUpload (Maybe Text)

Specifies presentational information for the object.

createMultipartUpload_contentLanguage :: Lens' CreateMultipartUpload (Maybe Text)

The language the content is in.

createMultipartUpload_sSEKMSEncryptionContext :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

createMultipartUpload_metadata :: Lens' CreateMultipartUpload (HashMap Text Text)

A map of metadata to store with the object in S3.

createMultipartUpload_sSECustomerKeyMD5 :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

createMultipartUpload_sSEKMSKeyId :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the ID of the symmetric customer managed key to use for object encryption. All GET and PUT requests for an object protected by Amazon Web Services KMS will fail if not made via SSL or using SigV4. For information about configuring using any of the officially supported Amazon Web Services SDKs and Amazon Web Services CLI, see Specifying the Signature Version in Request Authentication in the Amazon S3 User Guide.

createMultipartUpload_contentEncoding :: Lens' CreateMultipartUpload (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

createMultipartUpload_storageClass :: Lens' CreateMultipartUpload (Maybe StorageClass)

By default, Amazon S3 uses the STANDARD Storage Class to store newly created objects. The STANDARD storage class provides high durability and high availability. Depending on performance needs, you can specify a different Storage Class. Amazon S3 on Outposts only uses the OUTPOSTS Storage Class. For more information, see Storage Classes in the Amazon S3 User Guide.

createMultipartUpload_bucketKeyEnabled :: Lens' CreateMultipartUpload (Maybe Bool)

Specifies whether Amazon S3 should use an S3 Bucket Key for object encryption with server-side encryption using AWS KMS (SSE-KMS). Setting this header to true causes Amazon S3 to use an S3 Bucket Key for object encryption with SSE-KMS.

Specifying this header with an object action doesn’t affect bucket-level settings for S3 Bucket Key.

createMultipartUpload_grantWriteACP :: Lens' CreateMultipartUpload (Maybe Text)

Allows grantee to write the ACL for the applicable object.

This action is not supported by Amazon S3 on Outposts.

createMultipartUpload_serverSideEncryption :: Lens' CreateMultipartUpload (Maybe ServerSideEncryption)

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

createMultipartUpload_acl :: Lens' CreateMultipartUpload (Maybe ObjectCannedACL)

The canned ACL to apply to the object.

This action is not supported by Amazon S3 on Outposts.

createMultipartUpload_sSECustomerKey :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

createMultipartUpload_sSECustomerAlgorithm :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the algorithm to use to when encrypting the object (for example, AES256).

createMultipartUpload_grantReadACP :: Lens' CreateMultipartUpload (Maybe Text)

Allows grantee to read the object ACL.

This action is not supported by Amazon S3 on Outposts.

createMultipartUpload_objectLockLegalHoldStatus :: Lens' CreateMultipartUpload (Maybe ObjectLockLegalHoldStatus)

Specifies whether you want to apply a Legal Hold to the uploaded object.

createMultipartUpload_requestPayer :: Lens' CreateMultipartUpload (Maybe RequestPayer)

Undocumented member.

createMultipartUpload_cacheControl :: Lens' CreateMultipartUpload (Maybe Text)

Specifies caching behavior along the request/reply chain.

createMultipartUpload_expires :: Lens' CreateMultipartUpload (Maybe UTCTime)

The date and time at which the object is no longer cacheable.

createMultipartUpload_objectLockMode :: Lens' CreateMultipartUpload (Maybe ObjectLockMode)

Specifies the Object Lock mode that you want to apply to the uploaded object.

createMultipartUpload_objectLockRetainUntilDate :: Lens' CreateMultipartUpload (Maybe UTCTime)

Specifies the date and time when you want the Object Lock to expire.

createMultipartUpload_tagging :: Lens' CreateMultipartUpload (Maybe Text)

The tag-set for the object. The tag-set must be encoded as URL Query parameters.

createMultipartUpload_grantFullControl :: Lens' CreateMultipartUpload (Maybe Text)

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

This action is not supported by Amazon S3 on Outposts.

createMultipartUpload_bucket :: Lens' CreateMultipartUpload BucketName

The name of the bucket to which to initiate the upload

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

createMultipartUpload_key :: Lens' CreateMultipartUpload ObjectKey

Object key for which the multipart upload is to be initiated.

Destructuring the Response

data CreateMultipartUploadResponse

See: newCreateMultipartUploadResponse smart constructor.

Constructors

CreateMultipartUploadResponse'

Fields

key :: Maybe ObjectKey

Object key for which the multipart upload was initiated.

requestCharged :: Maybe RequestCharged

abortDate :: Maybe ISO8601

If the bucket has a lifecycle rule configured with an action to abort incomplete multipart uploads and the prefix in the lifecycle rule matches the object name in the request, the response includes this header. The header indicates when the initiated multipart upload becomes eligible for an abort operation. For more information, see Aborting Incomplete Multipart Uploads Using a Bucket Lifecycle Policy.

The response also includes the x-amz-abort-rule-id header that provides the ID of the lifecycle configuration rule that defines this action.

uploadId :: Maybe Text

ID for the initiated multipart upload.

sSEKMSEncryptionContext :: Maybe (Sensitive Text)

If present, specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

sSECustomerKeyMD5 :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

sSEKMSKeyId :: Maybe (Sensitive Text)

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

bucketKeyEnabled :: Maybe Bool

Indicates whether the multipart upload uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

abortRuleId :: Maybe Text

This header is returned along with the x-amz-abort-date header. It identifies the applicable lifecycle configuration rule that defines the action to abort incomplete multipart uploads.

serverSideEncryption :: Maybe ServerSideEncryption

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

bucket :: Maybe BucketName

The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

sSECustomerAlgorithm :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

httpStatus :: Int

The response's http status code.

newCreateMultipartUploadResponse

Arguments

:: Int
$sel:httpStatus:CreateMultipartUploadResponse'

-> CreateMultipartUploadResponse

Create a value of CreateMultipartUploadResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:CreateMultipartUpload', createMultipartUploadResponse_key - Object key for which the multipart upload was initiated.

$sel:requestCharged:CreateMultipartUploadResponse', createMultipartUploadResponse_requestCharged - Undocumented member.

$sel:abortDate:CreateMultipartUploadResponse', createMultipartUploadResponse_abortDate - If the bucket has a lifecycle rule configured with an action to abort incomplete multipart uploads and the prefix in the lifecycle rule matches the object name in the request, the response includes this header. The header indicates when the initiated multipart upload becomes eligible for an abort operation. For more information, see Aborting Incomplete Multipart Uploads Using a Bucket Lifecycle Policy.

The response also includes the x-amz-abort-rule-id header that provides the ID of the lifecycle configuration rule that defines this action.

$sel:uploadId:CreateMultipartUploadResponse', createMultipartUploadResponse_uploadId - ID for the initiated multipart upload.

$sel:sSEKMSEncryptionContext:CreateMultipartUpload', createMultipartUploadResponse_sSEKMSEncryptionContext - If present, specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

$sel:sSECustomerKeyMD5:CreateMultipartUpload', createMultipartUploadResponse_sSECustomerKeyMD5 - If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

$sel:sSEKMSKeyId:CreateMultipartUpload', createMultipartUploadResponse_sSEKMSKeyId - If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

$sel:bucketKeyEnabled:CreateMultipartUpload', createMultipartUploadResponse_bucketKeyEnabled - Indicates whether the multipart upload uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

$sel:abortRuleId:CreateMultipartUploadResponse', createMultipartUploadResponse_abortRuleId - This header is returned along with the x-amz-abort-date header. It identifies the applicable lifecycle configuration rule that defines the action to abort incomplete multipart uploads.

$sel:serverSideEncryption:CreateMultipartUpload', createMultipartUploadResponse_serverSideEncryption - The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

$sel:bucket:CreateMultipartUpload', createMultipartUploadResponse_bucket - The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

$sel:sSECustomerAlgorithm:CreateMultipartUpload', createMultipartUploadResponse_sSECustomerAlgorithm - If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

$sel:httpStatus:CreateMultipartUploadResponse', createMultipartUploadResponse_httpStatus - The response's http status code.

Response Lenses

createMultipartUploadResponse_key :: Lens' CreateMultipartUploadResponse (Maybe ObjectKey)

Object key for which the multipart upload was initiated.

createMultipartUploadResponse_requestCharged :: Lens' CreateMultipartUploadResponse (Maybe RequestCharged)

Undocumented member.

createMultipartUploadResponse_abortDate :: Lens' CreateMultipartUploadResponse (Maybe UTCTime)

If the bucket has a lifecycle rule configured with an action to abort incomplete multipart uploads and the prefix in the lifecycle rule matches the object name in the request, the response includes this header. The header indicates when the initiated multipart upload becomes eligible for an abort operation. For more information, see Aborting Incomplete Multipart Uploads Using a Bucket Lifecycle Policy.

The response also includes the x-amz-abort-rule-id header that provides the ID of the lifecycle configuration rule that defines this action.

createMultipartUploadResponse_uploadId :: Lens' CreateMultipartUploadResponse (Maybe Text)

ID for the initiated multipart upload.

createMultipartUploadResponse_sSEKMSEncryptionContext :: Lens' CreateMultipartUploadResponse (Maybe Text)

If present, specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

createMultipartUploadResponse_sSECustomerKeyMD5 :: Lens' CreateMultipartUploadResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

createMultipartUploadResponse_sSEKMSKeyId :: Lens' CreateMultipartUploadResponse (Maybe Text)

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

createMultipartUploadResponse_bucketKeyEnabled :: Lens' CreateMultipartUploadResponse (Maybe Bool)

Indicates whether the multipart upload uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

createMultipartUploadResponse_abortRuleId :: Lens' CreateMultipartUploadResponse (Maybe Text)

This header is returned along with the x-amz-abort-date header. It identifies the applicable lifecycle configuration rule that defines the action to abort incomplete multipart uploads.

createMultipartUploadResponse_serverSideEncryption :: Lens' CreateMultipartUploadResponse (Maybe ServerSideEncryption)

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

createMultipartUploadResponse_bucket :: Lens' CreateMultipartUploadResponse (Maybe BucketName)

The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

createMultipartUploadResponse_sSECustomerAlgorithm :: Lens' CreateMultipartUploadResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

createMultipartUploadResponse_httpStatus :: Lens' CreateMultipartUploadResponse Int

The response's http status code.