Network.AWS.S3.DeleteBucketAnalyticsConfiguration

Description

Deletes an analytics configuration for the bucket (specified by the analytics configuration ID).

To use this operation, you must have permissions to perform the s3:PutAnalyticsConfiguration action. The bucket owner has this permission by default. The bucket owner can grant this permission to others. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to Your Amazon S3 Resources.

For information about the Amazon S3 analytics feature, see Amazon S3 Analytics – Storage Class Analysis.

The following operations are related to DeleteBucketAnalyticsConfiguration:

GetBucketAnalyticsConfiguration

ListBucketAnalyticsConfigurations

PutBucketAnalyticsConfiguration

Creating a Request

data DeleteBucketAnalyticsConfiguration

See: newDeleteBucketAnalyticsConfiguration smart constructor.

Constructors

DeleteBucketAnalyticsConfiguration'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The name of the bucket from which an analytics configuration is deleted.

id :: Text

The ID that identifies the analytics configuration.

newDeleteBucketAnalyticsConfiguration

Arguments

:: BucketName
$sel:bucket:DeleteBucketAnalyticsConfiguration'

-> Text
$sel:id:DeleteBucketAnalyticsConfiguration'

-> DeleteBucketAnalyticsConfiguration

Create a value of DeleteBucketAnalyticsConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:DeleteBucketAnalyticsConfiguration', deleteBucketAnalyticsConfiguration_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:DeleteBucketAnalyticsConfiguration', deleteBucketAnalyticsConfiguration_bucket - The name of the bucket from which an analytics configuration is deleted.

$sel:id:DeleteBucketAnalyticsConfiguration', deleteBucketAnalyticsConfiguration_id - The ID that identifies the analytics configuration.

Request Lenses

deleteBucketAnalyticsConfiguration_expectedBucketOwner :: Lens' DeleteBucketAnalyticsConfiguration (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

deleteBucketAnalyticsConfiguration_bucket :: Lens' DeleteBucketAnalyticsConfiguration BucketName

The name of the bucket from which an analytics configuration is deleted.

deleteBucketAnalyticsConfiguration_id :: Lens' DeleteBucketAnalyticsConfiguration Text

The ID that identifies the analytics configuration.

Destructuring the Response

data DeleteBucketAnalyticsConfigurationResponse

See: newDeleteBucketAnalyticsConfigurationResponse smart constructor.

Constructors

DeleteBucketAnalyticsConfigurationResponse'

newDeleteBucketAnalyticsConfigurationResponse :: DeleteBucketAnalyticsConfigurationResponse

Create a value of DeleteBucketAnalyticsConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.