Network.AWS.S3.DeleteBucketIntelligentTieringConfiguration

Description

Deletes the S3 Intelligent-Tiering configuration from the specified bucket.

The S3 Intelligent-Tiering storage class is designed to optimize storage costs by automatically moving data to the most cost-effective storage access tier, without performance impact or operational overhead. S3 Intelligent-Tiering delivers automatic cost savings in two low latency and high throughput access tiers. For data that can be accessed asynchronously, you can choose to activate automatic archiving capabilities within the S3 Intelligent-Tiering storage class.

The S3 Intelligent-Tiering storage class is the ideal storage class for data with unknown, changing, or unpredictable access patterns, independent of object size or retention period. If the size of an object is less than 128 KB, it is not eligible for auto-tiering. Smaller objects can be stored, but they are always charged at the Frequent Access tier rates in the S3 Intelligent-Tiering storage class.

For more information, see Storage class for automatically optimizing frequently and infrequently accessed objects.

Operations related to DeleteBucketIntelligentTieringConfiguration include:

GetBucketIntelligentTieringConfiguration

PutBucketIntelligentTieringConfiguration

ListBucketIntelligentTieringConfigurations

Creating a Request

data DeleteBucketIntelligentTieringConfiguration

See: newDeleteBucketIntelligentTieringConfiguration smart constructor.

Constructors

DeleteBucketIntelligentTieringConfiguration'

Fields

bucket :: BucketName

The name of the Amazon S3 bucket whose configuration you want to modify or retrieve.

id :: Text

The ID used to identify the S3 Intelligent-Tiering configuration.

newDeleteBucketIntelligentTieringConfiguration

Arguments

:: BucketName
$sel:bucket:DeleteBucketIntelligentTieringConfiguration'

-> Text
$sel:id:DeleteBucketIntelligentTieringConfiguration'

-> DeleteBucketIntelligentTieringConfiguration

Create a value of DeleteBucketIntelligentTieringConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucket:DeleteBucketIntelligentTieringConfiguration', deleteBucketIntelligentTieringConfiguration_bucket - The name of the Amazon S3 bucket whose configuration you want to modify or retrieve.

$sel:id:DeleteBucketIntelligentTieringConfiguration', deleteBucketIntelligentTieringConfiguration_id - The ID used to identify the S3 Intelligent-Tiering configuration.

Request Lenses

deleteBucketIntelligentTieringConfiguration_bucket :: Lens' DeleteBucketIntelligentTieringConfiguration BucketName

The name of the Amazon S3 bucket whose configuration you want to modify or retrieve.

deleteBucketIntelligentTieringConfiguration_id :: Lens' DeleteBucketIntelligentTieringConfiguration Text

The ID used to identify the S3 Intelligent-Tiering configuration.

Destructuring the Response

data DeleteBucketIntelligentTieringConfigurationResponse

See: newDeleteBucketIntelligentTieringConfigurationResponse smart constructor.

Constructors

DeleteBucketIntelligentTieringConfigurationResponse'

newDeleteBucketIntelligentTieringConfigurationResponse :: DeleteBucketIntelligentTieringConfigurationResponse

Create a value of DeleteBucketIntelligentTieringConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.