Network.AWS.S3.DeleteBucketWebsite

Description

This action removes the website configuration for a bucket. Amazon S3 returns a 200 OK response upon successfully deleting a website configuration on the specified bucket. You will get a 200 OK response if the website configuration you are trying to delete does not exist on the bucket. Amazon S3 returns a 404 response if the bucket specified in the request does not exist.

This DELETE action requires the S3:DeleteBucketWebsite permission. By default, only the bucket owner can delete the website configuration attached to a bucket. However, bucket owners can grant other users permission to delete the website configuration by writing a bucket policy granting them the S3:DeleteBucketWebsite permission.

For more information about hosting websites, see Hosting Websites on Amazon S3.

The following operations are related to DeleteBucketWebsite:

GetBucketWebsite

PutBucketWebsite

Creating a Request

data DeleteBucketWebsite

See: newDeleteBucketWebsite smart constructor.

Constructors

DeleteBucketWebsite'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The bucket name for which you want to remove the website configuration.

newDeleteBucketWebsite

Arguments

:: BucketName
$sel:bucket:DeleteBucketWebsite'

-> DeleteBucketWebsite

Create a value of DeleteBucketWebsite with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:DeleteBucketWebsite', deleteBucketWebsite_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:DeleteBucketWebsite', deleteBucketWebsite_bucket - The bucket name for which you want to remove the website configuration.

Request Lenses

deleteBucketWebsite_expectedBucketOwner :: Lens' DeleteBucketWebsite (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

deleteBucketWebsite_bucket :: Lens' DeleteBucketWebsite BucketName

The bucket name for which you want to remove the website configuration.

Destructuring the Response

data DeleteBucketWebsiteResponse

See: newDeleteBucketWebsiteResponse smart constructor.

Constructors

DeleteBucketWebsiteResponse'

newDeleteBucketWebsiteResponse :: DeleteBucketWebsiteResponse

Create a value of DeleteBucketWebsiteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.