Network.AWS.S3.DeletePublicAccessBlock

Description

Removes the PublicAccessBlock configuration for an Amazon S3 bucket. To use this operation, you must have the s3:PutBucketPublicAccessBlock permission. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to Your Amazon S3 Resources.

The following operations are related to DeletePublicAccessBlock:

Using Amazon S3 Block Public Access

GetPublicAccessBlock

PutPublicAccessBlock

GetBucketPolicyStatus

Creating a Request

data DeletePublicAccessBlock

See: newDeletePublicAccessBlock smart constructor.

Constructors

DeletePublicAccessBlock'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The Amazon S3 bucket whose PublicAccessBlock configuration you want to delete.

newDeletePublicAccessBlock

Arguments

:: BucketName
$sel:bucket:DeletePublicAccessBlock'

-> DeletePublicAccessBlock

Create a value of DeletePublicAccessBlock with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:DeletePublicAccessBlock', deletePublicAccessBlock_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:DeletePublicAccessBlock', deletePublicAccessBlock_bucket - The Amazon S3 bucket whose PublicAccessBlock configuration you want to delete.

Request Lenses

deletePublicAccessBlock_expectedBucketOwner :: Lens' DeletePublicAccessBlock (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

deletePublicAccessBlock_bucket :: Lens' DeletePublicAccessBlock BucketName

The Amazon S3 bucket whose PublicAccessBlock configuration you want to delete.

Destructuring the Response

data DeletePublicAccessBlockResponse

See: newDeletePublicAccessBlockResponse smart constructor.

Constructors

DeletePublicAccessBlockResponse'

newDeletePublicAccessBlockResponse :: DeletePublicAccessBlockResponse

Create a value of DeletePublicAccessBlockResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.