Network.AWS.S3.Encryption.Envelope

Description

Documentation

data V1Envelope

Constructors

V1Envelope

Fields

_v1Key :: !ByteString

x-amz-key: Content encrypting key (cek) in encrypted form, base64 encoded. The cek is randomly generated per S3 object, and is always an AES 256-bit key. The corresponding cipher is always AESCBCPKCS5Padding.

_v1IV :: !(IV AES256)

x-amz-iv: Randomly generated IV (per S3 object), base64 encoded.

_v1Description :: !Description

x-amz-matdesc: Customer provided material description in JSON (UTF8) format.

newV1 :: MonadIO m => (ByteString -> IO ByteString) -> Description -> m Envelope

decodeV1 :: MonadResource m => (ByteString -> IO ByteString) -> [(CI Text, Text)] -> m Envelope

data V2Envelope

Constructors

V2Envelope

Fields

_v2Key :: !ByteString

x-amz-key-v2: CEK in key wrapped form. This is necessary so that the S3 encryption client that doesn't recognize the v2 format will not mistakenly decrypt S3 object encrypted in v2 format.

_v2IV :: !(IV AES256)

x-amz-iv: Randomly generated IV (per S3 object), base64 encoded.

_v2CEKAlgorithm :: !ContentAlgorithm

x-amz-cek-alg: Content encryption algorithm used. Supported values: AESGCMNoPadding, AESCBCPKCS5Padding Default to AESCBCPKCS5Padding if this key is absent.

Supported values: AESWrap, RSAECBOAEPWithSHA-256AndMGF1Padding, kms No standard key wrapping is used if this meta information is absent Always set to kms if KMS is used for client-side encryption

_v2WrapAlgorithm :: !WrappingAlgorithm

x-amz-wrap-alg: Key wrapping algorithm used.

_v2Description :: !Description

x-amz-matdesc: Customer provided material description in JSON format. Used to identify the client-side master key. For KMS client side encryption, the KMS Customer Master Key ID is stored as part of the material description, x-amz-matdesc, under the key-name kms_cmk_id@.

newV2 :: MonadResource m => Text -> Env -> Description -> m Envelope

decodeV2 :: MonadResource m => Env -> [(CI Text, Text)] -> Description -> m Envelope

data Envelope

Constructors

V1 AES256 V1Envelope

V2 AES256 V2Envelope

toMetadata :: Envelope -> [(CI ByteString, ByteString)]

newEnvelope :: MonadResource m => Key -> Env -> m Envelope

decodeEnvelope :: MonadResource m => Key -> Env -> [(CI Text, Text)] -> m Envelope

fromMetadata :: MonadResource m => Key -> Env -> HashMap Text Text -> m Envelope

aesKeySize :: Int

aesBlockSize :: Int

bodyEncrypt :: Envelope -> RequestBody -> RequestBody

bodyDecrypt :: Envelope -> ResponseBody -> ResponseBody

aesCbc :: Monad m => IV AES256 -> (IV AES256 -> ByteString -> (IV AES256, ByteString)) -> (IV AES256 -> ByteString -> ByteString) -> ConduitT ByteString ByteString m ()

rsaEncrypt :: KeyPair -> ByteString -> IO ByteString

rsaDecrypt :: KeyPair -> ByteString -> IO ByteString

getCipher :: Envelope -> (AES256, IV AES256)

createCipher :: (MonadIO m, ByteArray a, Cipher b) => a -> m b

createIV :: (MonadIO m, BlockCipher a) => ByteString -> m (IV a)

plaintext :: MonadIO m => DecryptResponse -> m ByteString

(.&) :: (MonadIO m, FromText a) => [(CI Text, Text)] -> CI Text -> m a

hoistEither :: MonadIO m => Either EncryptionError a -> m a

throwIO :: MonadIO m => EncryptionError -> m a