Network.AWS.S3.Encryption.Types

Description

Documentation

data EncryptionError

An error thrown when performing encryption or decryption.

Constructors

CipherFailure CryptoError
Error initialising an AES cipher from a secret key.

PubKeyFailure Error
Failure performing asymmetric encryption/decryption.

IVInvalid ByteString
Failure creating an IV from some bytes.

EnvelopeMissing (CI Text)
Required envelope field missing.

EnvelopeInvalid (CI Text) String
Error parsing envelope.

PlaintextUnavailable
KMS error when retrieving decrypted plaintext.

class AsEncryptionError r where

Minimal complete definition

_EncryptionError

Methods

_EncryptionError :: Prism' r EncryptionError

_CipherFailure :: Prism' r CryptoError

_PubKeyFailure :: Prism' r Error

_IVInvalid :: Prism' r ByteString

_EnvelopeMissing :: Prism' r (CI Text)

_EnvelopeInvalid :: Prism' r (CI Text, String)

_PlaintextUnavailable :: Prism' r ()

data ContentAlgorithm

Constructors

AES_CBC_PKCS5Padding
AESCBCPKCS5Padding

data WrappingAlgorithm

Constructors

KMSWrap
Key Management Service.

data Location

Constructors

Metadata

Discard

newtype Ext

An instructions file extension.

Constructors

Ext Text

defaultExtension :: Ext

Defaults to .instruction

appendExtension :: Ext -> ObjectKey -> ObjectKey

newtype Description

A key material description. This is attached in plaintext to the metadata, and will be logged using CloudTrail. For KMS decryption any supplemental material description is merged with the description stored on the object during decryption.

Constructors

Description

Fields

fromDescription :: HashMap Text Text

data Key

The key used for encryption and decryption.

Constructors

Symmetric AES256 Description

Asymmetric KeyPair Description

KMS Text Description

description :: Lens' Key Description

Modify the material description of a key.

See: Description.