Network.AWS.S3.GetBucketAccelerateConfiguration

Description

This implementation of the GET action uses the accelerate subresource to return the Transfer Acceleration state of a bucket, which is either Enabled or Suspended. Amazon S3 Transfer Acceleration is a bucket-level feature that enables you to perform faster data transfers to and from Amazon S3.

To use this operation, you must have permission to perform the s3:GetAccelerateConfiguration action. The bucket owner has this permission by default. The bucket owner can grant this permission to others. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to your Amazon S3 Resources in the Amazon S3 User Guide.

You set the Transfer Acceleration state of an existing bucket to Enabled or Suspended by using the PutBucketAccelerateConfiguration operation.

A GET accelerate request does not return a state value for a bucket that has no transfer acceleration state. A bucket has no Transfer Acceleration state if a state has never been set on the bucket.

For more information about transfer acceleration, see Transfer Acceleration in the Amazon S3 User Guide.

Related Resources

PutBucketAccelerateConfiguration

Creating a Request

data GetBucketAccelerateConfiguration

See: newGetBucketAccelerateConfiguration smart constructor.

Constructors

GetBucketAccelerateConfiguration'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The name of the bucket for which the accelerate configuration is retrieved.

newGetBucketAccelerateConfiguration

Arguments

:: BucketName
$sel:bucket:GetBucketAccelerateConfiguration'

-> GetBucketAccelerateConfiguration

Create a value of GetBucketAccelerateConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:GetBucketAccelerateConfiguration', getBucketAccelerateConfiguration_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:GetBucketAccelerateConfiguration', getBucketAccelerateConfiguration_bucket - The name of the bucket for which the accelerate configuration is retrieved.

Request Lenses

getBucketAccelerateConfiguration_expectedBucketOwner :: Lens' GetBucketAccelerateConfiguration (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

getBucketAccelerateConfiguration_bucket :: Lens' GetBucketAccelerateConfiguration BucketName

The name of the bucket for which the accelerate configuration is retrieved.

Destructuring the Response

data GetBucketAccelerateConfigurationResponse

See: newGetBucketAccelerateConfigurationResponse smart constructor.

Constructors

GetBucketAccelerateConfigurationResponse'

Fields

status :: Maybe BucketAccelerateStatus

The accelerate configuration of the bucket.

httpStatus :: Int

The response's http status code.

newGetBucketAccelerateConfigurationResponse

Arguments

:: Int
$sel:httpStatus:GetBucketAccelerateConfigurationResponse'

-> GetBucketAccelerateConfigurationResponse

Create a value of GetBucketAccelerateConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetBucketAccelerateConfigurationResponse', getBucketAccelerateConfigurationResponse_status - The accelerate configuration of the bucket.

$sel:httpStatus:GetBucketAccelerateConfigurationResponse', getBucketAccelerateConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

getBucketAccelerateConfigurationResponse_status :: Lens' GetBucketAccelerateConfigurationResponse (Maybe BucketAccelerateStatus)

The accelerate configuration of the bucket.

getBucketAccelerateConfigurationResponse_httpStatus :: Lens' GetBucketAccelerateConfigurationResponse Int

The response's http status code.