Network.AWS.S3.GetBucketAcl

Description

This implementation of the GET action uses the acl subresource to return the access control list (ACL) of a bucket. To use GET to return the ACL of the bucket, you must have READ_ACP access to the bucket. If READ_ACP permission is granted to the anonymous user, you can return the ACL of the bucket without using an authorization header.

Related Resources

ListObjects

Creating a Request

data GetBucketAcl

See: newGetBucketAcl smart constructor.

Constructors

GetBucketAcl'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

Specifies the S3 bucket whose ACL is being requested.

newGetBucketAcl

Arguments

:: BucketName
$sel:bucket:GetBucketAcl'

-> GetBucketAcl

Create a value of GetBucketAcl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:GetBucketAcl', getBucketAcl_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:GetBucketAcl', getBucketAcl_bucket - Specifies the S3 bucket whose ACL is being requested.

Request Lenses

getBucketAcl_expectedBucketOwner :: Lens' GetBucketAcl (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

getBucketAcl_bucket :: Lens' GetBucketAcl BucketName

Specifies the S3 bucket whose ACL is being requested.

Destructuring the Response

data GetBucketAclResponse

See: newGetBucketAclResponse smart constructor.

Constructors

GetBucketAclResponse'

Fields

grants :: Maybe [Grant]

A list of grants.

owner :: Maybe Owner

Container for the bucket owner's display name and ID.

httpStatus :: Int

The response's http status code.

newGetBucketAclResponse

Arguments

:: Int
$sel:httpStatus:GetBucketAclResponse'

-> GetBucketAclResponse

Create a value of GetBucketAclResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:grants:GetBucketAclResponse', getBucketAclResponse_grants - A list of grants.

$sel:owner:GetBucketAclResponse', getBucketAclResponse_owner - Container for the bucket owner's display name and ID.

$sel:httpStatus:GetBucketAclResponse', getBucketAclResponse_httpStatus - The response's http status code.

Response Lenses

getBucketAclResponse_grants :: Lens' GetBucketAclResponse (Maybe [Grant])

A list of grants.

getBucketAclResponse_owner :: Lens' GetBucketAclResponse (Maybe Owner)

Container for the bucket owner's display name and ID.

getBucketAclResponse_httpStatus :: Lens' GetBucketAclResponse Int

The response's http status code.