Network.AWS.S3.GetBucketAnalyticsConfiguration

Description

This implementation of the GET action returns an analytics configuration (identified by the analytics configuration ID) from the bucket.

To use this operation, you must have permissions to perform the s3:GetAnalyticsConfiguration action. The bucket owner has this permission by default. The bucket owner can grant this permission to others. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to Your Amazon S3 Resources in the Amazon S3 User Guide.

For information about Amazon S3 analytics feature, see Amazon S3 Analytics – Storage Class Analysis in the Amazon S3 User Guide.

Related Resources

DeleteBucketAnalyticsConfiguration

ListBucketAnalyticsConfigurations

PutBucketAnalyticsConfiguration

Creating a Request

data GetBucketAnalyticsConfiguration

See: newGetBucketAnalyticsConfiguration smart constructor.

Constructors

GetBucketAnalyticsConfiguration'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The name of the bucket from which an analytics configuration is retrieved.

id :: Text

The ID that identifies the analytics configuration.

newGetBucketAnalyticsConfiguration

Arguments

:: BucketName
$sel:bucket:GetBucketAnalyticsConfiguration'

-> Text
$sel:id:GetBucketAnalyticsConfiguration'

-> GetBucketAnalyticsConfiguration

Create a value of GetBucketAnalyticsConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:GetBucketAnalyticsConfiguration', getBucketAnalyticsConfiguration_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:GetBucketAnalyticsConfiguration', getBucketAnalyticsConfiguration_bucket - The name of the bucket from which an analytics configuration is retrieved.

$sel:id:GetBucketAnalyticsConfiguration', getBucketAnalyticsConfiguration_id - The ID that identifies the analytics configuration.

Request Lenses

getBucketAnalyticsConfiguration_expectedBucketOwner :: Lens' GetBucketAnalyticsConfiguration (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

getBucketAnalyticsConfiguration_bucket :: Lens' GetBucketAnalyticsConfiguration BucketName

The name of the bucket from which an analytics configuration is retrieved.

getBucketAnalyticsConfiguration_id :: Lens' GetBucketAnalyticsConfiguration Text

The ID that identifies the analytics configuration.

Destructuring the Response

data GetBucketAnalyticsConfigurationResponse

See: newGetBucketAnalyticsConfigurationResponse smart constructor.

Constructors

GetBucketAnalyticsConfigurationResponse'

Fields

analyticsConfiguration :: Maybe AnalyticsConfiguration

The configuration and any analyses for the analytics filter.

httpStatus :: Int

The response's http status code.

newGetBucketAnalyticsConfigurationResponse

Arguments

:: Int
$sel:httpStatus:GetBucketAnalyticsConfigurationResponse'

-> GetBucketAnalyticsConfigurationResponse

Create a value of GetBucketAnalyticsConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:analyticsConfiguration:GetBucketAnalyticsConfigurationResponse', getBucketAnalyticsConfigurationResponse_analyticsConfiguration - The configuration and any analyses for the analytics filter.

$sel:httpStatus:GetBucketAnalyticsConfigurationResponse', getBucketAnalyticsConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

getBucketAnalyticsConfigurationResponse_analyticsConfiguration :: Lens' GetBucketAnalyticsConfigurationResponse (Maybe AnalyticsConfiguration)

The configuration and any analyses for the analytics filter.

getBucketAnalyticsConfigurationResponse_httpStatus :: Lens' GetBucketAnalyticsConfigurationResponse Int

The response's http status code.