Network.AWS.S3.GetBucketEncryption

Description

Returns the default encryption configuration for an Amazon S3 bucket. If the bucket does not have a default encryption configuration, GetBucketEncryption returns ServerSideEncryptionConfigurationNotFoundError.

For information about the Amazon S3 default encryption feature, see Amazon S3 Default Bucket Encryption.

To use this operation, you must have permission to perform the s3:GetEncryptionConfiguration action. The bucket owner has this permission by default. The bucket owner can grant this permission to others. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to Your Amazon S3 Resources.

The following operations are related to GetBucketEncryption:

PutBucketEncryption

DeleteBucketEncryption

Creating a Request

data GetBucketEncryption

See: newGetBucketEncryption smart constructor.

Constructors

GetBucketEncryption'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The name of the bucket from which the server-side encryption configuration is retrieved.

newGetBucketEncryption

Arguments

:: BucketName
$sel:bucket:GetBucketEncryption'

-> GetBucketEncryption

Create a value of GetBucketEncryption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:GetBucketEncryption', getBucketEncryption_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:GetBucketEncryption', getBucketEncryption_bucket - The name of the bucket from which the server-side encryption configuration is retrieved.

Request Lenses

getBucketEncryption_expectedBucketOwner :: Lens' GetBucketEncryption (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

getBucketEncryption_bucket :: Lens' GetBucketEncryption BucketName

The name of the bucket from which the server-side encryption configuration is retrieved.

Destructuring the Response

data GetBucketEncryptionResponse

See: newGetBucketEncryptionResponse smart constructor.

Constructors

GetBucketEncryptionResponse'

Fields

serverSideEncryptionConfiguration :: Maybe ServerSideEncryptionConfiguration

httpStatus :: Int

The response's http status code.

newGetBucketEncryptionResponse

Arguments

:: Int
$sel:httpStatus:GetBucketEncryptionResponse'

-> GetBucketEncryptionResponse

Create a value of GetBucketEncryptionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serverSideEncryptionConfiguration:GetBucketEncryptionResponse', getBucketEncryptionResponse_serverSideEncryptionConfiguration - Undocumented member.

$sel:httpStatus:GetBucketEncryptionResponse', getBucketEncryptionResponse_httpStatus - The response's http status code.

Response Lenses

getBucketEncryptionResponse_serverSideEncryptionConfiguration :: Lens' GetBucketEncryptionResponse (Maybe ServerSideEncryptionConfiguration)

Undocumented member.

getBucketEncryptionResponse_httpStatus :: Lens' GetBucketEncryptionResponse Int

The response's http status code.