Network.AWS.S3.GetBucketLifecycleConfiguration

Description

Bucket lifecycle configuration now supports specifying a lifecycle rule using an object key name prefix, one or more object tags, or a combination of both. Accordingly, this section describes the latest API. The response describes the new filter element that you can use to specify a filter to select a subset of objects to which the rule applies. If you are using a previous version of the lifecycle configuration, it still works. For the earlier action, see GetBucketLifecycle.

Returns the lifecycle configuration information set on the bucket. For information about lifecycle configuration, see Object Lifecycle Management.

To use this operation, you must have permission to perform the s3:GetLifecycleConfiguration action. The bucket owner has this permission, by default. The bucket owner can grant this permission to others. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to Your Amazon S3 Resources.

GetBucketLifecycleConfiguration has the following special error:

Error code: NoSuchLifecycleConfiguration

Description: The lifecycle configuration does not exist.

HTTP Status Code: 404 Not Found

SOAP Fault Code Prefix: Client

The following operations are related to GetBucketLifecycleConfiguration:

GetBucketLifecycle

PutBucketLifecycle

DeleteBucketLifecycle

Creating a Request

data GetBucketLifecycleConfiguration

See: newGetBucketLifecycleConfiguration smart constructor.

Constructors

GetBucketLifecycleConfiguration'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The name of the bucket for which to get the lifecycle information.

newGetBucketLifecycleConfiguration

Arguments

:: BucketName
$sel:bucket:GetBucketLifecycleConfiguration'

-> GetBucketLifecycleConfiguration

Create a value of GetBucketLifecycleConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:GetBucketLifecycleConfiguration', getBucketLifecycleConfiguration_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:GetBucketLifecycleConfiguration', getBucketLifecycleConfiguration_bucket - The name of the bucket for which to get the lifecycle information.

Request Lenses

getBucketLifecycleConfiguration_expectedBucketOwner :: Lens' GetBucketLifecycleConfiguration (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

getBucketLifecycleConfiguration_bucket :: Lens' GetBucketLifecycleConfiguration BucketName

The name of the bucket for which to get the lifecycle information.

Destructuring the Response

data GetBucketLifecycleConfigurationResponse

See: newGetBucketLifecycleConfigurationResponse smart constructor.

Constructors

GetBucketLifecycleConfigurationResponse'

Fields

rules :: Maybe [LifecycleRule]

Container for a lifecycle rule.

httpStatus :: Int

The response's http status code.

newGetBucketLifecycleConfigurationResponse

Arguments

:: Int
$sel:httpStatus:GetBucketLifecycleConfigurationResponse'

-> GetBucketLifecycleConfigurationResponse

Create a value of GetBucketLifecycleConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rules:GetBucketLifecycleConfigurationResponse', getBucketLifecycleConfigurationResponse_rules - Container for a lifecycle rule.

$sel:httpStatus:GetBucketLifecycleConfigurationResponse', getBucketLifecycleConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

getBucketLifecycleConfigurationResponse_rules :: Lens' GetBucketLifecycleConfigurationResponse (Maybe [LifecycleRule])

Container for a lifecycle rule.

getBucketLifecycleConfigurationResponse_httpStatus :: Lens' GetBucketLifecycleConfigurationResponse Int

The response's http status code.