Network.AWS.S3.GetBucketMetricsConfiguration

Description

Gets a metrics configuration (specified by the metrics configuration ID) from the bucket. Note that this doesn't include the daily storage metrics.

To use this operation, you must have permissions to perform the s3:GetMetricsConfiguration action. The bucket owner has this permission by default. The bucket owner can grant this permission to others. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to Your Amazon S3 Resources.

For information about CloudWatch request metrics for Amazon S3, see Monitoring Metrics with Amazon CloudWatch.

The following operations are related to GetBucketMetricsConfiguration:

PutBucketMetricsConfiguration

DeleteBucketMetricsConfiguration

ListBucketMetricsConfigurations

Monitoring Metrics with Amazon CloudWatch

Creating a Request

data GetBucketMetricsConfiguration

See: newGetBucketMetricsConfiguration smart constructor.

Constructors

GetBucketMetricsConfiguration'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The name of the bucket containing the metrics configuration to retrieve.

id :: Text

The ID used to identify the metrics configuration.

newGetBucketMetricsConfiguration

Arguments

:: BucketName
$sel:bucket:GetBucketMetricsConfiguration'

-> Text
$sel:id:GetBucketMetricsConfiguration'

-> GetBucketMetricsConfiguration

Create a value of GetBucketMetricsConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:GetBucketMetricsConfiguration', getBucketMetricsConfiguration_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:GetBucketMetricsConfiguration', getBucketMetricsConfiguration_bucket - The name of the bucket containing the metrics configuration to retrieve.

$sel:id:GetBucketMetricsConfiguration', getBucketMetricsConfiguration_id - The ID used to identify the metrics configuration.

Request Lenses

getBucketMetricsConfiguration_expectedBucketOwner :: Lens' GetBucketMetricsConfiguration (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

getBucketMetricsConfiguration_bucket :: Lens' GetBucketMetricsConfiguration BucketName

The name of the bucket containing the metrics configuration to retrieve.

getBucketMetricsConfiguration_id :: Lens' GetBucketMetricsConfiguration Text

The ID used to identify the metrics configuration.

Destructuring the Response

data GetBucketMetricsConfigurationResponse

See: newGetBucketMetricsConfigurationResponse smart constructor.

Constructors

GetBucketMetricsConfigurationResponse'

Fields

metricsConfiguration :: Maybe MetricsConfiguration

Specifies the metrics configuration.

httpStatus :: Int

The response's http status code.

newGetBucketMetricsConfigurationResponse

Arguments

:: Int
$sel:httpStatus:GetBucketMetricsConfigurationResponse'

-> GetBucketMetricsConfigurationResponse

Create a value of GetBucketMetricsConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricsConfiguration:GetBucketMetricsConfigurationResponse', getBucketMetricsConfigurationResponse_metricsConfiguration - Specifies the metrics configuration.

$sel:httpStatus:GetBucketMetricsConfigurationResponse', getBucketMetricsConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

getBucketMetricsConfigurationResponse_metricsConfiguration :: Lens' GetBucketMetricsConfigurationResponse (Maybe MetricsConfiguration)

Specifies the metrics configuration.

getBucketMetricsConfigurationResponse_httpStatus :: Lens' GetBucketMetricsConfigurationResponse Int

The response's http status code.