Network.AWS.S3.GetBucketPolicy

Description

Returns the policy of a specified bucket. If you are using an identity other than the root user of the Amazon Web Services account that owns the bucket, the calling identity must have the GetBucketPolicy permissions on the specified bucket and belong to the bucket owner's account in order to use this operation.

If you don't have GetBucketPolicy permissions, Amazon S3 returns a 403 Access Denied error. If you have the correct permissions, but you're not using an identity that belongs to the bucket owner's account, Amazon S3 returns a 405 Method Not Allowed error.

As a security precaution, the root user of the Amazon Web Services account that owns a bucket can always use this operation, even if the policy explicitly denies the root user the ability to perform this action.

For more information about bucket policies, see Using Bucket Policies and User Policies.

The following action is related to GetBucketPolicy:

GetObject

Creating a Request

data GetBucketPolicy

See: newGetBucketPolicy smart constructor.

Constructors

GetBucketPolicy'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The bucket name for which to get the bucket policy.

newGetBucketPolicy

Arguments

:: BucketName
$sel:bucket:GetBucketPolicy'

-> GetBucketPolicy

Create a value of GetBucketPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:GetBucketPolicy', getBucketPolicy_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:GetBucketPolicy', getBucketPolicy_bucket - The bucket name for which to get the bucket policy.

Request Lenses

getBucketPolicy_expectedBucketOwner :: Lens' GetBucketPolicy (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

getBucketPolicy_bucket :: Lens' GetBucketPolicy BucketName

The bucket name for which to get the bucket policy.

Destructuring the Response

data GetBucketPolicyResponse

See: newGetBucketPolicyResponse smart constructor.

Constructors

GetBucketPolicyResponse'

Fields

httpStatus :: Int

The response's http status code.

policy :: ByteString

The bucket policy as a JSON document.

newGetBucketPolicyResponse

Arguments

:: Int
$sel:httpStatus:GetBucketPolicyResponse'

-> ByteString
$sel:policy:GetBucketPolicyResponse'

-> GetBucketPolicyResponse

Create a value of GetBucketPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetBucketPolicyResponse', getBucketPolicyResponse_httpStatus - The response's http status code.

$sel:policy:GetBucketPolicyResponse', getBucketPolicyResponse_policy - The bucket policy as a JSON document.

Response Lenses

getBucketPolicyResponse_httpStatus :: Lens' GetBucketPolicyResponse Int

The response's http status code.

getBucketPolicyResponse_policy :: Lens' GetBucketPolicyResponse ByteString

The bucket policy as a JSON document.