Network.AWS.S3.GetBucketReplication

Description

Returns the replication configuration of a bucket.

It can take a while to propagate the put or delete a replication configuration to all Amazon S3 systems. Therefore, a get request soon after put or delete can return a wrong result.

For information about replication configuration, see Replication in the Amazon S3 User Guide.

This action requires permissions for the s3:GetReplicationConfiguration action. For more information about permissions, see Using Bucket Policies and User Policies.

If you include the Filter element in a replication configuration, you must also include the DeleteMarkerReplication and Priority elements. The response also returns those elements.

For information about GetBucketReplication errors, see List of replication-related error codes

The following operations are related to GetBucketReplication:

PutBucketReplication

DeleteBucketReplication

Creating a Request

data GetBucketReplication

See: newGetBucketReplication smart constructor.

Constructors

GetBucketReplication'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The bucket name for which to get the replication information.

newGetBucketReplication

Arguments

:: BucketName
$sel:bucket:GetBucketReplication'

-> GetBucketReplication

Create a value of GetBucketReplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:GetBucketReplication', getBucketReplication_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:GetBucketReplication', getBucketReplication_bucket - The bucket name for which to get the replication information.

Request Lenses

getBucketReplication_expectedBucketOwner :: Lens' GetBucketReplication (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

getBucketReplication_bucket :: Lens' GetBucketReplication BucketName

The bucket name for which to get the replication information.

Destructuring the Response

data GetBucketReplicationResponse

See: newGetBucketReplicationResponse smart constructor.

Constructors

GetBucketReplicationResponse'

Fields

replicationConfiguration :: Maybe ReplicationConfiguration

httpStatus :: Int

The response's http status code.

newGetBucketReplicationResponse

Arguments

:: Int
$sel:httpStatus:GetBucketReplicationResponse'

-> GetBucketReplicationResponse

Create a value of GetBucketReplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationConfiguration:GetBucketReplicationResponse', getBucketReplicationResponse_replicationConfiguration - Undocumented member.

$sel:httpStatus:GetBucketReplicationResponse', getBucketReplicationResponse_httpStatus - The response's http status code.

Response Lenses

getBucketReplicationResponse_replicationConfiguration :: Lens' GetBucketReplicationResponse (Maybe ReplicationConfiguration)

Undocumented member.

getBucketReplicationResponse_httpStatus :: Lens' GetBucketReplicationResponse Int

The response's http status code.