Network.AWS.S3.GetBucketVersioning

Description

Returns the versioning state of a bucket.

To retrieve the versioning state of a bucket, you must be the bucket owner.

This implementation also returns the MFA Delete status of the versioning state. If the MFA Delete status is enabled, the bucket owner must use an authentication device to change the versioning state of the bucket.

The following operations are related to GetBucketVersioning:

GetObject

PutObject

DeleteObject

Creating a Request

data GetBucketVersioning

See: newGetBucketVersioning smart constructor.

Constructors

GetBucketVersioning'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The name of the bucket for which to get the versioning information.

newGetBucketVersioning

Arguments

:: BucketName
$sel:bucket:GetBucketVersioning'

-> GetBucketVersioning

Create a value of GetBucketVersioning with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:GetBucketVersioning', getBucketVersioning_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:GetBucketVersioning', getBucketVersioning_bucket - The name of the bucket for which to get the versioning information.

Request Lenses

getBucketVersioning_expectedBucketOwner :: Lens' GetBucketVersioning (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

getBucketVersioning_bucket :: Lens' GetBucketVersioning BucketName

The name of the bucket for which to get the versioning information.

Destructuring the Response

data GetBucketVersioningResponse

See: newGetBucketVersioningResponse smart constructor.

Constructors

GetBucketVersioningResponse'

Fields

status :: Maybe BucketVersioningStatus

The versioning state of the bucket.

mfaDelete :: Maybe MFADeleteStatus

Specifies whether MFA delete is enabled in the bucket versioning configuration. This element is only returned if the bucket has been configured with MFA delete. If the bucket has never been so configured, this element is not returned.

httpStatus :: Int

The response's http status code.

newGetBucketVersioningResponse

Arguments

:: Int
$sel:httpStatus:GetBucketVersioningResponse'

-> GetBucketVersioningResponse

Create a value of GetBucketVersioningResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetBucketVersioningResponse', getBucketVersioningResponse_status - The versioning state of the bucket.

$sel:mfaDelete:GetBucketVersioningResponse', getBucketVersioningResponse_mfaDelete - Specifies whether MFA delete is enabled in the bucket versioning configuration. This element is only returned if the bucket has been configured with MFA delete. If the bucket has never been so configured, this element is not returned.

$sel:httpStatus:GetBucketVersioningResponse', getBucketVersioningResponse_httpStatus - The response's http status code.

Response Lenses

getBucketVersioningResponse_status :: Lens' GetBucketVersioningResponse (Maybe BucketVersioningStatus)

The versioning state of the bucket.

getBucketVersioningResponse_mfaDelete :: Lens' GetBucketVersioningResponse (Maybe MFADeleteStatus)

Specifies whether MFA delete is enabled in the bucket versioning configuration. This element is only returned if the bucket has been configured with MFA delete. If the bucket has never been so configured, this element is not returned.

getBucketVersioningResponse_httpStatus :: Lens' GetBucketVersioningResponse Int

The response's http status code.