Network.AWS.S3.GetBucketWebsite

Description

Returns the website configuration for a bucket. To host website on Amazon S3, you can configure a bucket as website by adding a website configuration. For more information about hosting websites, see Hosting Websites on Amazon S3.

This GET action requires the S3:GetBucketWebsite permission. By default, only the bucket owner can read the bucket website configuration. However, bucket owners can allow other users to read the website configuration by writing a bucket policy granting them the S3:GetBucketWebsite permission.

The following operations are related to DeleteBucketWebsite:

DeleteBucketWebsite

PutBucketWebsite

Creating a Request

data GetBucketWebsite

See: newGetBucketWebsite smart constructor.

Constructors

GetBucketWebsite'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The bucket name for which to get the website configuration.

newGetBucketWebsite

Arguments

:: BucketName
$sel:bucket:GetBucketWebsite'

-> GetBucketWebsite

Create a value of GetBucketWebsite with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:GetBucketWebsite', getBucketWebsite_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:GetBucketWebsite', getBucketWebsite_bucket - The bucket name for which to get the website configuration.

Request Lenses

getBucketWebsite_expectedBucketOwner :: Lens' GetBucketWebsite (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

getBucketWebsite_bucket :: Lens' GetBucketWebsite BucketName

The bucket name for which to get the website configuration.

Destructuring the Response

data GetBucketWebsiteResponse

See: newGetBucketWebsiteResponse smart constructor.

Constructors

GetBucketWebsiteResponse'

Fields

errorDocument :: Maybe ErrorDocument

The object key name of the website error document to use for 4XX class errors.

indexDocument :: Maybe IndexDocument

The name of the index document for the website (for example index.html).

routingRules :: Maybe [RoutingRule]

Rules that define when a redirect is applied and the redirect behavior.

redirectAllRequestsTo :: Maybe RedirectAllRequestsTo

Specifies the redirect behavior of all requests to a website endpoint of an Amazon S3 bucket.

httpStatus :: Int

The response's http status code.

newGetBucketWebsiteResponse

Arguments

:: Int
$sel:httpStatus:GetBucketWebsiteResponse'

-> GetBucketWebsiteResponse

Create a value of GetBucketWebsiteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorDocument:GetBucketWebsiteResponse', getBucketWebsiteResponse_errorDocument - The object key name of the website error document to use for 4XX class errors.

$sel:indexDocument:GetBucketWebsiteResponse', getBucketWebsiteResponse_indexDocument - The name of the index document for the website (for example index.html).

$sel:routingRules:GetBucketWebsiteResponse', getBucketWebsiteResponse_routingRules - Rules that define when a redirect is applied and the redirect behavior.

$sel:redirectAllRequestsTo:GetBucketWebsiteResponse', getBucketWebsiteResponse_redirectAllRequestsTo - Specifies the redirect behavior of all requests to a website endpoint of an Amazon S3 bucket.

$sel:httpStatus:GetBucketWebsiteResponse', getBucketWebsiteResponse_httpStatus - The response's http status code.

Response Lenses

getBucketWebsiteResponse_errorDocument :: Lens' GetBucketWebsiteResponse (Maybe ErrorDocument)

The object key name of the website error document to use for 4XX class errors.

getBucketWebsiteResponse_indexDocument :: Lens' GetBucketWebsiteResponse (Maybe IndexDocument)

The name of the index document for the website (for example index.html).

getBucketWebsiteResponse_routingRules :: Lens' GetBucketWebsiteResponse (Maybe [RoutingRule])

Rules that define when a redirect is applied and the redirect behavior.

getBucketWebsiteResponse_redirectAllRequestsTo :: Lens' GetBucketWebsiteResponse (Maybe RedirectAllRequestsTo)

Specifies the redirect behavior of all requests to a website endpoint of an Amazon S3 bucket.

getBucketWebsiteResponse_httpStatus :: Lens' GetBucketWebsiteResponse Int

The response's http status code.