Network.AWS.S3.GetObject

Description

Retrieves objects from Amazon S3. To use GET, you must have READ access to the object. If you grant READ access to the anonymous user, you can return the object without using an authorization header.

An Amazon S3 bucket has no directory hierarchy such as you would find in a typical computer file system. You can, however, create a logical hierarchy by using object key names that imply a folder structure. For example, instead of naming an object sample.jpg, you can name it photos/2006/February/sample.jpg.

To get an object from such a logical hierarchy, specify the full key name for the object in the GET operation. For a virtual hosted-style request example, if you have the object photos/2006/February/sample.jpg, specify the resource as /photos/2006/February/sample.jpg. For a path-style request example, if you have the object photos/2006/February/sample.jpg in the bucket named examplebucket, specify the resource as /examplebucket/photos/2006/February/sample.jpg. For more information about request types, see HTTP Host Header Bucket Specification.

To distribute large files to many people, you can save bandwidth costs by using BitTorrent. For more information, see Amazon S3 Torrent. For more information about returning the ACL of an object, see GetObjectAcl.

If the object you are retrieving is stored in the S3 Glacier or S3 Glacier Deep Archive storage class, or S3 Intelligent-Tiering Archive or S3 Intelligent-Tiering Deep Archive tiers, before you can retrieve the object you must first restore a copy using RestoreObject. Otherwise, this action returns an InvalidObjectStateError error. For information about restoring archived objects, see Restoring Archived Objects.

Encryption request headers, like x-amz-server-side-encryption, should not be sent for GET requests if your object uses server-side encryption with KMS keys (SSE-KMS) or server-side encryption with Amazon S3–managed encryption keys (SSE-S3). If your object does use these types of keys, you’ll get an HTTP 400 BadRequest error.

If you encrypt an object by using server-side encryption with customer-provided encryption keys (SSE-C) when you store the object in Amazon S3, then when you GET the object, you must use the following headers:

x-amz-server-side-encryption-customer-algorithm

x-amz-server-side-encryption-customer-key

x-amz-server-side-encryption-customer-key-MD5

For more information about SSE-C, see Server-Side Encryption (Using Customer-Provided Encryption Keys).

Assuming you have the relevant permission to read object tags, the response also returns the x-amz-tagging-count header that provides the count of number of tags associated with the object. You can use GetObjectTagging to retrieve the tag set associated with an object.

Permissions

You need the relevant read object (or version) permission for this operation. For more information, see Specifying Permissions in a Policy. If the object you request does not exist, the error Amazon S3 returns depends on whether you also have the s3:ListBucket permission.

If you have the s3:ListBucket permission on the bucket, Amazon S3 will return an HTTP status code 404 ("no such key") error.

If you don’t have the s3:ListBucket permission, Amazon S3 will return an HTTP status code 403 ("access denied") error.

Versioning

By default, the GET action returns the current version of an object. To return a different version, use the versionId subresource.

You need the s3:GetObjectVersion permission to access a specific version of an object.

If the current version of the object is a delete marker, Amazon S3 behaves as if the object was deleted and includes x-amz-delete-marker: true in the response.

For more information about versioning, see PutBucketVersioning.

Overriding Response Header Values

There are times when you want to override certain response header values in a GET response. For example, you might override the Content-Disposition response header value in your GET request.

You can override values for a set of response headers using the following query parameters. These response header values are sent only on a successful request, that is, when status code 200 OK is returned. The set of headers you can override using these parameters is a subset of the headers that Amazon S3 accepts when you create an object. The response headers that you can override for the GET response are Content-Type, Content-Language, Expires, Cache-Control, Content-Disposition, and Content-Encoding. To override these header values in the GET response, you use the following request parameters.

You must sign the request, either using an Authorization header or a presigned URL, when using these parameters. They cannot be used with an unsigned (anonymous) request.

response-content-type

response-content-language

response-expires

response-cache-control

response-content-disposition

response-content-encoding

Additional Considerations about Request Headers

If both of the If-Match and If-Unmodified-Since headers are present in the request as follows: If-Match condition evaluates to true, and; If-Unmodified-Since condition evaluates to false; then, S3 returns 200 OK and the data requested.

If both of the If-None-Match and If-Modified-Since headers are present in the request as follows: If-None-Match condition evaluates to false, and; If-Modified-Since condition evaluates to true; then, S3 returns 304 Not Modified response code.

For more information about conditional requests, see RFC 7232.

The following operations are related to GetObject:

ListBuckets

GetObjectAcl

Creating a Request

data GetObject

See: newGetObject smart constructor.

Constructors

GetObject'

Fields

ifUnmodifiedSince :: Maybe ISO8601

Return the object only if it has not been modified since the specified time, otherwise return a 412 (precondition failed).

range :: Maybe Text

Downloads the specified range bytes of an object. For more information about the HTTP Range header, see https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.35.

Amazon S3 doesn't support retrieving multiple ranges of data per GET request.

ifModifiedSince :: Maybe ISO8601

Return the object only if it has been modified since the specified time, otherwise return a 304 (not modified).

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

responseExpires :: Maybe ISO8601

Sets the Expires header of the response.

sSECustomerKeyMD5 :: Maybe Text

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

versionId :: Maybe ObjectVersionId

VersionId used to reference a specific version of the object.

ifMatch :: Maybe Text

Return the object only if its entity tag (ETag) is the same as the one specified, otherwise return a 412 (precondition failed).

partNumber :: Maybe Int

Part number of the object being read. This is a positive integer between 1 and 10,000. Effectively performs a 'ranged' GET request for the part specified. Useful for downloading just a part of an object.

responseContentEncoding :: Maybe Text

Sets the Content-Encoding header of the response.

sSECustomerKey :: Maybe (Sensitive Text)

Specifies the customer-provided encryption key for Amazon S3 used to encrypt the data. This value is used to decrypt the object when recovering it and must match the one used when storing the data. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

sSECustomerAlgorithm :: Maybe Text

Specifies the algorithm to use to when decrypting the object (for example, AES256).

ifNoneMatch :: Maybe Text

Return the object only if its entity tag (ETag) is different from the one specified, otherwise return a 304 (not modified).

requestPayer :: Maybe RequestPayer

responseContentLanguage :: Maybe Text

Sets the Content-Language header of the response.

responseCacheControl :: Maybe Text

Sets the Cache-Control header of the response.

responseContentDisposition :: Maybe Text

Sets the Content-Disposition header of the response

responseContentType :: Maybe Text

Sets the Content-Type header of the response.

bucket :: BucketName

The bucket name containing the object.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using an Object Lambda access point the hostname takes the form AccessPointName-AccountId.s3-object-lambda.Region.amazonaws.com.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

key :: ObjectKey

Key of the object to get.

newGetObject

Arguments

:: BucketName
$sel:bucket:GetObject'

-> ObjectKey
$sel:key:GetObject'

-> GetObject

Create a value of GetObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ifUnmodifiedSince:GetObject', getObject_ifUnmodifiedSince - Return the object only if it has not been modified since the specified time, otherwise return a 412 (precondition failed).

$sel:range:GetObject', getObject_range - Downloads the specified range bytes of an object. For more information about the HTTP Range header, see https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.35.

Amazon S3 doesn't support retrieving multiple ranges of data per GET request.

$sel:ifModifiedSince:GetObject', getObject_ifModifiedSince - Return the object only if it has been modified since the specified time, otherwise return a 304 (not modified).

$sel:expectedBucketOwner:GetObject', getObject_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:responseExpires:GetObject', getObject_responseExpires - Sets the Expires header of the response.

$sel:sSECustomerKeyMD5:GetObject', getObject_sSECustomerKeyMD5 - Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

$sel:versionId:GetObject', getObject_versionId - VersionId used to reference a specific version of the object.

$sel:ifMatch:GetObject', getObject_ifMatch - Return the object only if its entity tag (ETag) is the same as the one specified, otherwise return a 412 (precondition failed).

$sel:partNumber:GetObject', getObject_partNumber - Part number of the object being read. This is a positive integer between 1 and 10,000. Effectively performs a 'ranged' GET request for the part specified. Useful for downloading just a part of an object.

$sel:responseContentEncoding:GetObject', getObject_responseContentEncoding - Sets the Content-Encoding header of the response.

$sel:sSECustomerKey:GetObject', getObject_sSECustomerKey - Specifies the customer-provided encryption key for Amazon S3 used to encrypt the data. This value is used to decrypt the object when recovering it and must match the one used when storing the data. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

$sel:sSECustomerAlgorithm:GetObject', getObject_sSECustomerAlgorithm - Specifies the algorithm to use to when decrypting the object (for example, AES256).

$sel:ifNoneMatch:GetObject', getObject_ifNoneMatch - Return the object only if its entity tag (ETag) is different from the one specified, otherwise return a 304 (not modified).

$sel:requestPayer:GetObject', getObject_requestPayer - Undocumented member.

$sel:responseContentLanguage:GetObject', getObject_responseContentLanguage - Sets the Content-Language header of the response.

$sel:responseCacheControl:GetObject', getObject_responseCacheControl - Sets the Cache-Control header of the response.

$sel:responseContentDisposition:GetObject', getObject_responseContentDisposition - Sets the Content-Disposition header of the response

$sel:responseContentType:GetObject', getObject_responseContentType - Sets the Content-Type header of the response.

$sel:bucket:GetObject', getObject_bucket - The bucket name containing the object.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using an Object Lambda access point the hostname takes the form AccessPointName-AccountId.s3-object-lambda.Region.amazonaws.com.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

$sel:key:GetObject', getObject_key - Key of the object to get.

Request Lenses

getObject_ifUnmodifiedSince :: Lens' GetObject (Maybe UTCTime)

Return the object only if it has not been modified since the specified time, otherwise return a 412 (precondition failed).

getObject_range :: Lens' GetObject (Maybe Text)

Downloads the specified range bytes of an object. For more information about the HTTP Range header, see https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.35.

Amazon S3 doesn't support retrieving multiple ranges of data per GET request.

getObject_ifModifiedSince :: Lens' GetObject (Maybe UTCTime)

Return the object only if it has been modified since the specified time, otherwise return a 304 (not modified).

getObject_expectedBucketOwner :: Lens' GetObject (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

getObject_responseExpires :: Lens' GetObject (Maybe UTCTime)

Sets the Expires header of the response.

getObject_sSECustomerKeyMD5 :: Lens' GetObject (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

getObject_versionId :: Lens' GetObject (Maybe ObjectVersionId)

VersionId used to reference a specific version of the object.

getObject_ifMatch :: Lens' GetObject (Maybe Text)

Return the object only if its entity tag (ETag) is the same as the one specified, otherwise return a 412 (precondition failed).

getObject_partNumber :: Lens' GetObject (Maybe Int)

Part number of the object being read. This is a positive integer between 1 and 10,000. Effectively performs a 'ranged' GET request for the part specified. Useful for downloading just a part of an object.

getObject_responseContentEncoding :: Lens' GetObject (Maybe Text)

Sets the Content-Encoding header of the response.

getObject_sSECustomerKey :: Lens' GetObject (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 used to encrypt the data. This value is used to decrypt the object when recovering it and must match the one used when storing the data. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

getObject_sSECustomerAlgorithm :: Lens' GetObject (Maybe Text)

Specifies the algorithm to use to when decrypting the object (for example, AES256).

getObject_ifNoneMatch :: Lens' GetObject (Maybe Text)

Return the object only if its entity tag (ETag) is different from the one specified, otherwise return a 304 (not modified).

getObject_requestPayer :: Lens' GetObject (Maybe RequestPayer)

Undocumented member.

getObject_responseContentLanguage :: Lens' GetObject (Maybe Text)

Sets the Content-Language header of the response.

getObject_responseCacheControl :: Lens' GetObject (Maybe Text)

Sets the Cache-Control header of the response.

getObject_responseContentDisposition :: Lens' GetObject (Maybe Text)

Sets the Content-Disposition header of the response

getObject_responseContentType :: Lens' GetObject (Maybe Text)

Sets the Content-Type header of the response.

getObject_bucket :: Lens' GetObject BucketName

The bucket name containing the object.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using an Object Lambda access point the hostname takes the form AccessPointName-AccountId.s3-object-lambda.Region.amazonaws.com.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

getObject_key :: Lens' GetObject ObjectKey

Key of the object to get.

Destructuring the Response

data GetObjectResponse

See: newGetObjectResponse smart constructor.

Constructors

GetObjectResponse'

Fields

partsCount :: Maybe Int

The count of parts this object has.

requestCharged :: Maybe RequestCharged

eTag :: Maybe ETag

An ETag is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

websiteRedirectLocation :: Maybe Text

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

tagCount :: Maybe Int

The number of tags, if any, on the object.

contentType :: Maybe Text

A standard MIME type describing the format of the object data.

contentDisposition :: Maybe Text

Specifies presentational information for the object.

contentRange :: Maybe Text

The portion of the object returned in the response.

deleteMarker :: Maybe Bool

Specifies whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

contentLanguage :: Maybe Text

The language the content is in.

expiration :: Maybe Text

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL encoded.

replicationStatus :: Maybe ReplicationStatus

Amazon S3 can return this if your request involves a bucket that is either a source or destination in a replication rule.

contentLength :: Maybe Integer

Size of the body in bytes.

metadata :: HashMap Text Text

A map of metadata to store with the object in S3.

sSECustomerKeyMD5 :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

sSEKMSKeyId :: Maybe (Sensitive Text)

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

contentEncoding :: Maybe Text

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

versionId :: Maybe ObjectVersionId

Version of the object.

storageClass :: Maybe StorageClass

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

bucketKeyEnabled :: Maybe Bool

Indicates whether the object uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

acceptRanges :: Maybe Text

Indicates that a range of bytes was specified.

missingMeta :: Maybe Int

This is set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

serverSideEncryption :: Maybe ServerSideEncryption

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

sSECustomerAlgorithm :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

objectLockLegalHoldStatus :: Maybe ObjectLockLegalHoldStatus

Indicates whether this object has an active legal hold. This field is only returned if you have permission to view an object's legal hold status.

cacheControl :: Maybe Text

Specifies caching behavior along the request/reply chain.

lastModified :: Maybe ISO8601

Creation date of the object.

expires :: Maybe ISO8601

The date and time at which the object is no longer cacheable.

restore :: Maybe Text

Provides information about object restoration action and expiration time of the restored object copy.

objectLockMode :: Maybe ObjectLockMode

The Object Lock mode currently in place for this object.

objectLockRetainUntilDate :: Maybe ISO8601

The date and time when this object's Object Lock will expire.

httpStatus :: Int

The response's http status code.

body :: ResponseBody

Object data.

newGetObjectResponse

Arguments

:: Int
$sel:httpStatus:GetObjectResponse'

-> ResponseBody
$sel:body:GetObjectResponse'

-> GetObjectResponse

Create a value of GetObjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:partsCount:GetObjectResponse', getObjectResponse_partsCount - The count of parts this object has.

$sel:requestCharged:GetObjectResponse', getObjectResponse_requestCharged - Undocumented member.

$sel:eTag:GetObjectResponse', getObjectResponse_eTag - An ETag is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

$sel:websiteRedirectLocation:GetObjectResponse', getObjectResponse_websiteRedirectLocation - If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

$sel:tagCount:GetObjectResponse', getObjectResponse_tagCount - The number of tags, if any, on the object.

$sel:contentType:GetObjectResponse', getObjectResponse_contentType - A standard MIME type describing the format of the object data.

$sel:contentDisposition:GetObjectResponse', getObjectResponse_contentDisposition - Specifies presentational information for the object.

$sel:contentRange:GetObjectResponse', getObjectResponse_contentRange - The portion of the object returned in the response.

$sel:deleteMarker:GetObjectResponse', getObjectResponse_deleteMarker - Specifies whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

$sel:contentLanguage:GetObjectResponse', getObjectResponse_contentLanguage - The language the content is in.

$sel:expiration:GetObjectResponse', getObjectResponse_expiration - If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL encoded.

$sel:replicationStatus:GetObjectResponse', getObjectResponse_replicationStatus - Amazon S3 can return this if your request involves a bucket that is either a source or destination in a replication rule.

$sel:contentLength:GetObjectResponse', getObjectResponse_contentLength - Size of the body in bytes.

$sel:metadata:GetObjectResponse', getObjectResponse_metadata - A map of metadata to store with the object in S3.

$sel:sSECustomerKeyMD5:GetObject', getObjectResponse_sSECustomerKeyMD5 - If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

$sel:sSEKMSKeyId:GetObjectResponse', getObjectResponse_sSEKMSKeyId - If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

$sel:contentEncoding:GetObjectResponse', getObjectResponse_contentEncoding - Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

$sel:versionId:GetObject', getObjectResponse_versionId - Version of the object.

$sel:storageClass:GetObjectResponse', getObjectResponse_storageClass - Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

$sel:bucketKeyEnabled:GetObjectResponse', getObjectResponse_bucketKeyEnabled - Indicates whether the object uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

$sel:acceptRanges:GetObjectResponse', getObjectResponse_acceptRanges - Indicates that a range of bytes was specified.

$sel:missingMeta:GetObjectResponse', getObjectResponse_missingMeta - This is set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

$sel:serverSideEncryption:GetObjectResponse', getObjectResponse_serverSideEncryption - The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

$sel:sSECustomerAlgorithm:GetObject', getObjectResponse_sSECustomerAlgorithm - If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

$sel:objectLockLegalHoldStatus:GetObjectResponse', getObjectResponse_objectLockLegalHoldStatus - Indicates whether this object has an active legal hold. This field is only returned if you have permission to view an object's legal hold status.

$sel:cacheControl:GetObjectResponse', getObjectResponse_cacheControl - Specifies caching behavior along the request/reply chain.

$sel:lastModified:GetObjectResponse', getObjectResponse_lastModified - Creation date of the object.

$sel:expires:GetObjectResponse', getObjectResponse_expires - The date and time at which the object is no longer cacheable.

$sel:restore:GetObjectResponse', getObjectResponse_restore - Provides information about object restoration action and expiration time of the restored object copy.

$sel:objectLockMode:GetObjectResponse', getObjectResponse_objectLockMode - The Object Lock mode currently in place for this object.

$sel:objectLockRetainUntilDate:GetObjectResponse', getObjectResponse_objectLockRetainUntilDate - The date and time when this object's Object Lock will expire.

$sel:httpStatus:GetObjectResponse', getObjectResponse_httpStatus - The response's http status code.

$sel:body:GetObjectResponse', getObjectResponse_body - Object data.

Response Lenses

getObjectResponse_partsCount :: Lens' GetObjectResponse (Maybe Int)

The count of parts this object has.

getObjectResponse_requestCharged :: Lens' GetObjectResponse (Maybe RequestCharged)

Undocumented member.

getObjectResponse_eTag :: Lens' GetObjectResponse (Maybe ETag)

An ETag is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

getObjectResponse_websiteRedirectLocation :: Lens' GetObjectResponse (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

getObjectResponse_tagCount :: Lens' GetObjectResponse (Maybe Int)

The number of tags, if any, on the object.

getObjectResponse_contentType :: Lens' GetObjectResponse (Maybe Text)

A standard MIME type describing the format of the object data.

getObjectResponse_contentDisposition :: Lens' GetObjectResponse (Maybe Text)

Specifies presentational information for the object.

getObjectResponse_contentRange :: Lens' GetObjectResponse (Maybe Text)

The portion of the object returned in the response.

getObjectResponse_deleteMarker :: Lens' GetObjectResponse (Maybe Bool)

Specifies whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

getObjectResponse_contentLanguage :: Lens' GetObjectResponse (Maybe Text)

The language the content is in.

getObjectResponse_expiration :: Lens' GetObjectResponse (Maybe Text)

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL encoded.

getObjectResponse_replicationStatus :: Lens' GetObjectResponse (Maybe ReplicationStatus)

Amazon S3 can return this if your request involves a bucket that is either a source or destination in a replication rule.

getObjectResponse_contentLength :: Lens' GetObjectResponse (Maybe Integer)

Size of the body in bytes.

getObjectResponse_metadata :: Lens' GetObjectResponse (HashMap Text Text)

A map of metadata to store with the object in S3.

getObjectResponse_sSECustomerKeyMD5 :: Lens' GetObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

getObjectResponse_sSEKMSKeyId :: Lens' GetObjectResponse (Maybe Text)

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

getObjectResponse_contentEncoding :: Lens' GetObjectResponse (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

getObjectResponse_versionId :: Lens' GetObjectResponse (Maybe ObjectVersionId)

Version of the object.

getObjectResponse_storageClass :: Lens' GetObjectResponse (Maybe StorageClass)

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

getObjectResponse_bucketKeyEnabled :: Lens' GetObjectResponse (Maybe Bool)

Indicates whether the object uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

getObjectResponse_acceptRanges :: Lens' GetObjectResponse (Maybe Text)

Indicates that a range of bytes was specified.

getObjectResponse_missingMeta :: Lens' GetObjectResponse (Maybe Int)

This is set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

getObjectResponse_serverSideEncryption :: Lens' GetObjectResponse (Maybe ServerSideEncryption)

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

getObjectResponse_sSECustomerAlgorithm :: Lens' GetObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

getObjectResponse_objectLockLegalHoldStatus :: Lens' GetObjectResponse (Maybe ObjectLockLegalHoldStatus)

Indicates whether this object has an active legal hold. This field is only returned if you have permission to view an object's legal hold status.

getObjectResponse_cacheControl :: Lens' GetObjectResponse (Maybe Text)

Specifies caching behavior along the request/reply chain.

getObjectResponse_lastModified :: Lens' GetObjectResponse (Maybe UTCTime)

Creation date of the object.

getObjectResponse_expires :: Lens' GetObjectResponse (Maybe UTCTime)

The date and time at which the object is no longer cacheable.

getObjectResponse_restore :: Lens' GetObjectResponse (Maybe Text)

Provides information about object restoration action and expiration time of the restored object copy.

getObjectResponse_objectLockMode :: Lens' GetObjectResponse (Maybe ObjectLockMode)

The Object Lock mode currently in place for this object.

getObjectResponse_objectLockRetainUntilDate :: Lens' GetObjectResponse (Maybe UTCTime)

The date and time when this object's Object Lock will expire.

getObjectResponse_httpStatus :: Lens' GetObjectResponse Int

The response's http status code.

getObjectResponse_body :: Lens' GetObjectResponse ResponseBody

Object data.