Network.AWS.S3.GetObjectAcl

Description

Returns the access control list (ACL) of an object. To use this operation, you must have READ_ACP access to the object.

This action is not supported by Amazon S3 on Outposts.

Versioning

By default, GET returns ACL information about the current version of an object. To return ACL information about a different version, use the versionId subresource.

The following operations are related to GetObjectAcl:

GetObject

DeleteObject

PutObject

Creating a Request

data GetObjectAcl

See: newGetObjectAcl smart constructor.

Constructors

GetObjectAcl'

Fields

versionId :: Maybe ObjectVersionId

VersionId used to reference a specific version of the object.

requestPayer :: Maybe RequestPayer

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The bucket name that contains the object for which to get the ACL information.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

key :: ObjectKey

The key of the object for which to get the ACL information.

newGetObjectAcl

Arguments

:: BucketName
$sel:bucket:GetObjectAcl'

-> ObjectKey
$sel:key:GetObjectAcl'

-> GetObjectAcl

Create a value of GetObjectAcl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:GetObjectAcl', getObjectAcl_versionId - VersionId used to reference a specific version of the object.

$sel:requestPayer:GetObjectAcl', getObjectAcl_requestPayer - Undocumented member.

$sel:expectedBucketOwner:GetObjectAcl', getObjectAcl_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:GetObjectAcl', getObjectAcl_bucket - The bucket name that contains the object for which to get the ACL information.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

$sel:key:GetObjectAcl', getObjectAcl_key - The key of the object for which to get the ACL information.

Request Lenses

getObjectAcl_versionId :: Lens' GetObjectAcl (Maybe ObjectVersionId)

VersionId used to reference a specific version of the object.

getObjectAcl_requestPayer :: Lens' GetObjectAcl (Maybe RequestPayer)

Undocumented member.

getObjectAcl_expectedBucketOwner :: Lens' GetObjectAcl (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

getObjectAcl_bucket :: Lens' GetObjectAcl BucketName

The bucket name that contains the object for which to get the ACL information.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

getObjectAcl_key :: Lens' GetObjectAcl ObjectKey

The key of the object for which to get the ACL information.

Destructuring the Response

data GetObjectAclResponse

See: newGetObjectAclResponse smart constructor.

Constructors

GetObjectAclResponse'

Fields

requestCharged :: Maybe RequestCharged

grants :: Maybe [Grant]

A list of grants.

owner :: Maybe Owner

Container for the bucket owner's display name and ID.

httpStatus :: Int

The response's http status code.

newGetObjectAclResponse

Arguments

:: Int
$sel:httpStatus:GetObjectAclResponse'

-> GetObjectAclResponse

Create a value of GetObjectAclResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestCharged:GetObjectAclResponse', getObjectAclResponse_requestCharged - Undocumented member.

$sel:grants:GetObjectAclResponse', getObjectAclResponse_grants - A list of grants.

$sel:owner:GetObjectAclResponse', getObjectAclResponse_owner - Container for the bucket owner's display name and ID.

$sel:httpStatus:GetObjectAclResponse', getObjectAclResponse_httpStatus - The response's http status code.

Response Lenses

getObjectAclResponse_requestCharged :: Lens' GetObjectAclResponse (Maybe RequestCharged)

Undocumented member.

getObjectAclResponse_grants :: Lens' GetObjectAclResponse (Maybe [Grant])

A list of grants.

getObjectAclResponse_owner :: Lens' GetObjectAclResponse (Maybe Owner)

Container for the bucket owner's display name and ID.

getObjectAclResponse_httpStatus :: Lens' GetObjectAclResponse Int

The response's http status code.