Network.AWS.S3.GetObjectTagging

Description

Returns the tag-set of an object. You send the GET request against the tagging subresource associated with the object.

To use this operation, you must have permission to perform the s3:GetObjectTagging action. By default, the GET action returns information about current version of an object. For a versioned bucket, you can have multiple versions of an object in your bucket. To retrieve tags of any other version, use the versionId query parameter. You also need permission for the s3:GetObjectVersionTagging action.

By default, the bucket owner has this permission and can grant this permission to others.

For information about the Amazon S3 object tagging feature, see Object Tagging.

The following action is related to GetObjectTagging:

PutObjectTagging

DeleteObjectTagging

Creating a Request

data GetObjectTagging

See: newGetObjectTagging smart constructor.

Constructors

GetObjectTagging'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

versionId :: Maybe ObjectVersionId

The versionId of the object for which to get the tagging information.

requestPayer :: Maybe RequestPayer

bucket :: BucketName

The bucket name containing the object for which to get the tagging information.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

key :: ObjectKey

Object key for which to get the tagging information.

newGetObjectTagging

Arguments

:: BucketName
$sel:bucket:GetObjectTagging'

-> ObjectKey
$sel:key:GetObjectTagging'

-> GetObjectTagging

Create a value of GetObjectTagging with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:GetObjectTagging', getObjectTagging_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:versionId:GetObjectTagging', getObjectTagging_versionId - The versionId of the object for which to get the tagging information.

$sel:requestPayer:GetObjectTagging', getObjectTagging_requestPayer - Undocumented member.

$sel:bucket:GetObjectTagging', getObjectTagging_bucket - The bucket name containing the object for which to get the tagging information.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

$sel:key:GetObjectTagging', getObjectTagging_key - Object key for which to get the tagging information.

Request Lenses

getObjectTagging_expectedBucketOwner :: Lens' GetObjectTagging (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

getObjectTagging_versionId :: Lens' GetObjectTagging (Maybe ObjectVersionId)

The versionId of the object for which to get the tagging information.

getObjectTagging_requestPayer :: Lens' GetObjectTagging (Maybe RequestPayer)

Undocumented member.

getObjectTagging_bucket :: Lens' GetObjectTagging BucketName

The bucket name containing the object for which to get the tagging information.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

getObjectTagging_key :: Lens' GetObjectTagging ObjectKey

Object key for which to get the tagging information.

Destructuring the Response

data GetObjectTaggingResponse

See: newGetObjectTaggingResponse smart constructor.

Constructors

GetObjectTaggingResponse'

Fields

versionId :: Maybe ObjectVersionId

The versionId of the object for which you got the tagging information.

httpStatus :: Int

The response's http status code.

tagSet :: [Tag]

Contains the tag set.

newGetObjectTaggingResponse

Arguments

:: Int
$sel:httpStatus:GetObjectTaggingResponse'

-> GetObjectTaggingResponse

Create a value of GetObjectTaggingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:GetObjectTagging', getObjectTaggingResponse_versionId - The versionId of the object for which you got the tagging information.

$sel:httpStatus:GetObjectTaggingResponse', getObjectTaggingResponse_httpStatus - The response's http status code.

$sel:tagSet:GetObjectTaggingResponse', getObjectTaggingResponse_tagSet - Contains the tag set.

Response Lenses

getObjectTaggingResponse_versionId :: Lens' GetObjectTaggingResponse (Maybe ObjectVersionId)

The versionId of the object for which you got the tagging information.

getObjectTaggingResponse_httpStatus :: Lens' GetObjectTaggingResponse Int

The response's http status code.

getObjectTaggingResponse_tagSet :: Lens' GetObjectTaggingResponse [Tag]

Contains the tag set.