Network.AWS.S3.GetObjectTorrent

Description

Returns torrent files from a bucket. BitTorrent can save you bandwidth when you're distributing large files. For more information about BitTorrent, see Using BitTorrent with Amazon S3.

You can get torrent only for objects that are less than 5 GB in size, and that are not encrypted using server-side encryption with a customer-provided encryption key.

To use GET, you must have READ access to the object.

This action is not supported by Amazon S3 on Outposts.

The following action is related to GetObjectTorrent:

GetObject

Creating a Request

data GetObjectTorrent

See: newGetObjectTorrent smart constructor.

Constructors

GetObjectTorrent'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

requestPayer :: Maybe RequestPayer

bucket :: BucketName

The name of the bucket containing the object for which to get the torrent files.

key :: ObjectKey

The object key for which to get the information.

newGetObjectTorrent

Arguments

:: BucketName
$sel:bucket:GetObjectTorrent'

-> ObjectKey
$sel:key:GetObjectTorrent'

-> GetObjectTorrent

Create a value of GetObjectTorrent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:GetObjectTorrent', getObjectTorrent_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:requestPayer:GetObjectTorrent', getObjectTorrent_requestPayer - Undocumented member.

$sel:bucket:GetObjectTorrent', getObjectTorrent_bucket - The name of the bucket containing the object for which to get the torrent files.

$sel:key:GetObjectTorrent', getObjectTorrent_key - The object key for which to get the information.

Request Lenses

getObjectTorrent_expectedBucketOwner :: Lens' GetObjectTorrent (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

getObjectTorrent_requestPayer :: Lens' GetObjectTorrent (Maybe RequestPayer)

Undocumented member.

getObjectTorrent_bucket :: Lens' GetObjectTorrent BucketName

The name of the bucket containing the object for which to get the torrent files.

getObjectTorrent_key :: Lens' GetObjectTorrent ObjectKey

The object key for which to get the information.

Destructuring the Response

data GetObjectTorrentResponse

See: newGetObjectTorrentResponse smart constructor.

Constructors

GetObjectTorrentResponse'

Fields

requestCharged :: Maybe RequestCharged

httpStatus :: Int

The response's http status code.

body :: ResponseBody

A Bencoded dictionary as defined by the BitTorrent specification

newGetObjectTorrentResponse

Arguments

:: Int
$sel:httpStatus:GetObjectTorrentResponse'

-> ResponseBody
$sel:body:GetObjectTorrentResponse'

-> GetObjectTorrentResponse

Create a value of GetObjectTorrentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestCharged:GetObjectTorrentResponse', getObjectTorrentResponse_requestCharged - Undocumented member.

$sel:httpStatus:GetObjectTorrentResponse', getObjectTorrentResponse_httpStatus - The response's http status code.

$sel:body:GetObjectTorrentResponse', getObjectTorrentResponse_body - A Bencoded dictionary as defined by the BitTorrent specification

Response Lenses

getObjectTorrentResponse_requestCharged :: Lens' GetObjectTorrentResponse (Maybe RequestCharged)

Undocumented member.

getObjectTorrentResponse_httpStatus :: Lens' GetObjectTorrentResponse Int

The response's http status code.

getObjectTorrentResponse_body :: Lens' GetObjectTorrentResponse ResponseBody

A Bencoded dictionary as defined by the BitTorrent specification