Network.AWS.S3.HeadObject

Description

The HEAD action retrieves metadata from an object without returning the object itself. This action is useful if you're only interested in an object's metadata. To use HEAD, you must have READ access to the object.

A HEAD request has the same options as a GET action on an object. The response is identical to the GET response except that there is no response body. Because of this, if the HEAD request generates an error, it returns a generic 404 Not Found or 403 Forbidden code. It is not possible to retrieve the exact exception beyond these error codes.

If you encrypt an object by using server-side encryption with customer-provided encryption keys (SSE-C) when you store the object in Amazon S3, then when you retrieve the metadata from the object, you must use the following headers:

x-amz-server-side-encryption-customer-algorithm

x-amz-server-side-encryption-customer-key

x-amz-server-side-encryption-customer-key-MD5

For more information about SSE-C, see Server-Side Encryption (Using Customer-Provided Encryption Keys).

Encryption request headers, like x-amz-server-side-encryption, should not be sent for GET requests if your object uses server-side encryption with KMS keys (SSE-KMS) or server-side encryption with Amazon S3–managed encryption keys (SSE-S3). If your object does use these types of keys, you’ll get an HTTP 400 BadRequest error.

The last modified property in this case is the creation date of the object.

Request headers are limited to 8 KB in size. For more information, see Common Request Headers.

Consider the following when using request headers:

Consideration 1 – If both of the If-Match and If-Unmodified-Since headers are present in the request as follows:

If-Match condition evaluates to true, and;

If-Unmodified-Since condition evaluates to false;

Then Amazon S3 returns 200 OK and the data requested.

Consideration 2 – If both of the If-None-Match and If-Modified-Since headers are present in the request as follows:

If-None-Match condition evaluates to false, and;

If-Modified-Since condition evaluates to true;

Then Amazon S3 returns the 304 Not Modified response code.

For more information about conditional requests, see RFC 7232.

Permissions

You need the relevant read object (or version) permission for this operation. For more information, see Specifying Permissions in a Policy. If the object you request does not exist, the error Amazon S3 returns depends on whether you also have the s3:ListBucket permission.

If you have the s3:ListBucket permission on the bucket, Amazon S3 returns an HTTP status code 404 ("no such key") error.

If you don’t have the s3:ListBucket permission, Amazon S3 returns an HTTP status code 403 ("access denied") error.

The following action is related to HeadObject:

GetObject

Creating a Request

data HeadObject

See: newHeadObject smart constructor.

Constructors

HeadObject'

Fields

ifUnmodifiedSince :: Maybe ISO8601

Return the object only if it has not been modified since the specified time, otherwise return a 412 (precondition failed).

range :: Maybe Text

Downloads the specified range bytes of an object. For more information about the HTTP Range header, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.35.

Amazon S3 doesn't support retrieving multiple ranges of data per GET request.

ifModifiedSince :: Maybe ISO8601

Return the object only if it has been modified since the specified time, otherwise return a 304 (not modified).

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

sSECustomerKeyMD5 :: Maybe Text

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

versionId :: Maybe ObjectVersionId

VersionId used to reference a specific version of the object.

ifMatch :: Maybe Text

Return the object only if its entity tag (ETag) is the same as the one specified, otherwise return a 412 (precondition failed).

partNumber :: Maybe Int

Part number of the object being read. This is a positive integer between 1 and 10,000. Effectively performs a 'ranged' HEAD request for the part specified. Useful querying about the size of the part and the number of parts in this object.

sSECustomerKey :: Maybe (Sensitive Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

sSECustomerAlgorithm :: Maybe Text

Specifies the algorithm to use to when encrypting the object (for example, AES256).

ifNoneMatch :: Maybe Text

Return the object only if its entity tag (ETag) is different from the one specified, otherwise return a 304 (not modified).

requestPayer :: Maybe RequestPayer

bucket :: BucketName

The name of the bucket containing the object.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

key :: ObjectKey

The object key.

newHeadObject

Arguments

:: BucketName
$sel:bucket:HeadObject'

-> ObjectKey
$sel:key:HeadObject'

-> HeadObject

Create a value of HeadObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ifUnmodifiedSince:HeadObject', headObject_ifUnmodifiedSince - Return the object only if it has not been modified since the specified time, otherwise return a 412 (precondition failed).

$sel:range:HeadObject', headObject_range - Downloads the specified range bytes of an object. For more information about the HTTP Range header, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.35.

Amazon S3 doesn't support retrieving multiple ranges of data per GET request.

$sel:ifModifiedSince:HeadObject', headObject_ifModifiedSince - Return the object only if it has been modified since the specified time, otherwise return a 304 (not modified).

$sel:expectedBucketOwner:HeadObject', headObject_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:sSECustomerKeyMD5:HeadObject', headObject_sSECustomerKeyMD5 - Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

$sel:versionId:HeadObject', headObject_versionId - VersionId used to reference a specific version of the object.

$sel:ifMatch:HeadObject', headObject_ifMatch - Return the object only if its entity tag (ETag) is the same as the one specified, otherwise return a 412 (precondition failed).

$sel:partNumber:HeadObject', headObject_partNumber - Part number of the object being read. This is a positive integer between 1 and 10,000. Effectively performs a 'ranged' HEAD request for the part specified. Useful querying about the size of the part and the number of parts in this object.

$sel:sSECustomerKey:HeadObject', headObject_sSECustomerKey - Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

$sel:sSECustomerAlgorithm:HeadObject', headObject_sSECustomerAlgorithm - Specifies the algorithm to use to when encrypting the object (for example, AES256).

$sel:ifNoneMatch:HeadObject', headObject_ifNoneMatch - Return the object only if its entity tag (ETag) is different from the one specified, otherwise return a 304 (not modified).

$sel:requestPayer:HeadObject', headObject_requestPayer - Undocumented member.

$sel:bucket:HeadObject', headObject_bucket - The name of the bucket containing the object.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

$sel:key:HeadObject', headObject_key - The object key.

Request Lenses

headObject_ifUnmodifiedSince :: Lens' HeadObject (Maybe UTCTime)

Return the object only if it has not been modified since the specified time, otherwise return a 412 (precondition failed).

headObject_range :: Lens' HeadObject (Maybe Text)

Downloads the specified range bytes of an object. For more information about the HTTP Range header, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.35.

Amazon S3 doesn't support retrieving multiple ranges of data per GET request.

headObject_ifModifiedSince :: Lens' HeadObject (Maybe UTCTime)

Return the object only if it has been modified since the specified time, otherwise return a 304 (not modified).

headObject_expectedBucketOwner :: Lens' HeadObject (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

headObject_sSECustomerKeyMD5 :: Lens' HeadObject (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

headObject_versionId :: Lens' HeadObject (Maybe ObjectVersionId)

VersionId used to reference a specific version of the object.

headObject_ifMatch :: Lens' HeadObject (Maybe Text)

Return the object only if its entity tag (ETag) is the same as the one specified, otherwise return a 412 (precondition failed).

headObject_partNumber :: Lens' HeadObject (Maybe Int)

Part number of the object being read. This is a positive integer between 1 and 10,000. Effectively performs a 'ranged' HEAD request for the part specified. Useful querying about the size of the part and the number of parts in this object.

headObject_sSECustomerKey :: Lens' HeadObject (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

headObject_sSECustomerAlgorithm :: Lens' HeadObject (Maybe Text)

Specifies the algorithm to use to when encrypting the object (for example, AES256).

headObject_ifNoneMatch :: Lens' HeadObject (Maybe Text)

Return the object only if its entity tag (ETag) is different from the one specified, otherwise return a 304 (not modified).

headObject_requestPayer :: Lens' HeadObject (Maybe RequestPayer)

Undocumented member.

headObject_bucket :: Lens' HeadObject BucketName

The name of the bucket containing the object.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

headObject_key :: Lens' HeadObject ObjectKey

The object key.

Destructuring the Response

data HeadObjectResponse

See: newHeadObjectResponse smart constructor.

Constructors

HeadObjectResponse'

Fields

partsCount :: Maybe Int

The count of parts this object has.

requestCharged :: Maybe RequestCharged

eTag :: Maybe ETag

An ETag is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

websiteRedirectLocation :: Maybe Text

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

contentType :: Maybe Text

A standard MIME type describing the format of the object data.

contentDisposition :: Maybe Text

Specifies presentational information for the object.

archiveStatus :: Maybe ArchiveStatus

The archive state of the head object.

deleteMarker :: Maybe Bool

Specifies whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

contentLanguage :: Maybe Text

The language the content is in.

expiration :: Maybe Text

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL encoded.

replicationStatus :: Maybe ReplicationStatus

Amazon S3 can return this header if your request involves a bucket that is either a source or a destination in a replication rule.

In replication, you have a source bucket on which you configure replication and destination bucket or buckets where Amazon S3 stores object replicas. When you request an object (GetObject) or object metadata (HeadObject) from these buckets, Amazon S3 will return the x-amz-replication-status header in the response as follows:

If requesting an object from the source bucket — Amazon S3 will return the x-amz-replication-status header if the object in your request is eligible for replication.

For example, suppose that in your replication configuration, you specify object prefix TaxDocs requesting Amazon S3 to replicate objects with key prefix TaxDocs. Any objects you upload with this key name prefix, for example TaxDocs/document1.pdf, are eligible for replication. For any object request with this key name prefix, Amazon S3 will return the x-amz-replication-status header with value PENDING, COMPLETED or FAILED indicating object replication status.

If requesting an object from a destination bucket — Amazon S3 will return the x-amz-replication-status header with value REPLICA if the object in your request is a replica that Amazon S3 created and there is no replica modification replication in progress.

When replicating objects to multiple destination buckets the x-amz-replication-status header acts differently. The header of the source object will only return a value of COMPLETED when replication is successful to all destinations. The header will remain at value PENDING until replication has completed for all destinations. If one or more destinations fails replication the header will return FAILED.

For more information, see Replication.

contentLength :: Maybe Integer

Size of the body in bytes.

metadata :: HashMap Text Text

A map of metadata to store with the object in S3.

sSECustomerKeyMD5 :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

sSEKMSKeyId :: Maybe (Sensitive Text)

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

contentEncoding :: Maybe Text

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

versionId :: Maybe ObjectVersionId

Version of the object.

storageClass :: Maybe StorageClass

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

For more information, see Storage Classes.

bucketKeyEnabled :: Maybe Bool

Indicates whether the object uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

acceptRanges :: Maybe Text

Indicates that a range of bytes was specified.

missingMeta :: Maybe Int

This is set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

serverSideEncryption :: Maybe ServerSideEncryption

If the object is stored using server-side encryption either with an Amazon Web Services KMS key or an Amazon S3-managed encryption key, the response includes this header with the value of the server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

sSECustomerAlgorithm :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

objectLockLegalHoldStatus :: Maybe ObjectLockLegalHoldStatus

Specifies whether a legal hold is in effect for this object. This header is only returned if the requester has the s3:GetObjectLegalHold permission. This header is not returned if the specified version of this object has never had a legal hold applied. For more information about S3 Object Lock, see Object Lock.

cacheControl :: Maybe Text

Specifies caching behavior along the request/reply chain.

lastModified :: Maybe ISO8601

Creation date of the object.

expires :: Maybe ISO8601

The date and time at which the object is no longer cacheable.

restore :: Maybe Text

If the object is an archived object (an object whose storage class is GLACIER), the response includes this header if either the archive restoration is in progress (see RestoreObject or an archive copy is already restored.

If an archive copy is already restored, the header value indicates when Amazon S3 is scheduled to delete the object copy. For example:

x-amz-restore: ongoing-request="false", expiry-date="Fri, 21 Dec 2012 00:00:00 GMT"

If the object restoration is in progress, the header returns the value ongoing-request="true".

For more information about archiving objects, see Transitioning Objects: General Considerations.

objectLockMode :: Maybe ObjectLockMode

The Object Lock mode, if any, that's in effect for this object. This header is only returned if the requester has the s3:GetObjectRetention permission. For more information about S3 Object Lock, see Object Lock.

objectLockRetainUntilDate :: Maybe ISO8601

The date and time when the Object Lock retention period expires. This header is only returned if the requester has the s3:GetObjectRetention permission.

httpStatus :: Int

The response's http status code.

newHeadObjectResponse

Arguments

:: Int
$sel:httpStatus:HeadObjectResponse'

-> HeadObjectResponse

Create a value of HeadObjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:partsCount:HeadObjectResponse', headObjectResponse_partsCount - The count of parts this object has.

$sel:requestCharged:HeadObjectResponse', headObjectResponse_requestCharged - Undocumented member.

$sel:eTag:HeadObjectResponse', headObjectResponse_eTag - An ETag is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

$sel:websiteRedirectLocation:HeadObjectResponse', headObjectResponse_websiteRedirectLocation - If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

$sel:contentType:HeadObjectResponse', headObjectResponse_contentType - A standard MIME type describing the format of the object data.

$sel:contentDisposition:HeadObjectResponse', headObjectResponse_contentDisposition - Specifies presentational information for the object.

$sel:archiveStatus:HeadObjectResponse', headObjectResponse_archiveStatus - The archive state of the head object.

$sel:deleteMarker:HeadObjectResponse', headObjectResponse_deleteMarker - Specifies whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

$sel:contentLanguage:HeadObjectResponse', headObjectResponse_contentLanguage - The language the content is in.

$sel:expiration:HeadObjectResponse', headObjectResponse_expiration - If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL encoded.

$sel:replicationStatus:HeadObjectResponse', headObjectResponse_replicationStatus - Amazon S3 can return this header if your request involves a bucket that is either a source or a destination in a replication rule.

In replication, you have a source bucket on which you configure replication and destination bucket or buckets where Amazon S3 stores object replicas. When you request an object (GetObject) or object metadata (HeadObject) from these buckets, Amazon S3 will return the x-amz-replication-status header in the response as follows:

If requesting an object from the source bucket — Amazon S3 will return the x-amz-replication-status header if the object in your request is eligible for replication.

For example, suppose that in your replication configuration, you specify object prefix TaxDocs requesting Amazon S3 to replicate objects with key prefix TaxDocs. Any objects you upload with this key name prefix, for example TaxDocs/document1.pdf, are eligible for replication. For any object request with this key name prefix, Amazon S3 will return the x-amz-replication-status header with value PENDING, COMPLETED or FAILED indicating object replication status.

If requesting an object from a destination bucket — Amazon S3 will return the x-amz-replication-status header with value REPLICA if the object in your request is a replica that Amazon S3 created and there is no replica modification replication in progress.

When replicating objects to multiple destination buckets the x-amz-replication-status header acts differently. The header of the source object will only return a value of COMPLETED when replication is successful to all destinations. The header will remain at value PENDING until replication has completed for all destinations. If one or more destinations fails replication the header will return FAILED.

For more information, see Replication.

$sel:contentLength:HeadObjectResponse', headObjectResponse_contentLength - Size of the body in bytes.

$sel:metadata:HeadObjectResponse', headObjectResponse_metadata - A map of metadata to store with the object in S3.

$sel:sSECustomerKeyMD5:HeadObject', headObjectResponse_sSECustomerKeyMD5 - If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

$sel:sSEKMSKeyId:HeadObjectResponse', headObjectResponse_sSEKMSKeyId - If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

$sel:contentEncoding:HeadObjectResponse', headObjectResponse_contentEncoding - Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

$sel:versionId:HeadObject', headObjectResponse_versionId - Version of the object.

$sel:storageClass:HeadObjectResponse', headObjectResponse_storageClass - Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

For more information, see Storage Classes.

$sel:bucketKeyEnabled:HeadObjectResponse', headObjectResponse_bucketKeyEnabled - Indicates whether the object uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

$sel:acceptRanges:HeadObjectResponse', headObjectResponse_acceptRanges - Indicates that a range of bytes was specified.

$sel:missingMeta:HeadObjectResponse', headObjectResponse_missingMeta - This is set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

$sel:serverSideEncryption:HeadObjectResponse', headObjectResponse_serverSideEncryption - If the object is stored using server-side encryption either with an Amazon Web Services KMS key or an Amazon S3-managed encryption key, the response includes this header with the value of the server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

$sel:sSECustomerAlgorithm:HeadObject', headObjectResponse_sSECustomerAlgorithm - If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

$sel:objectLockLegalHoldStatus:HeadObjectResponse', headObjectResponse_objectLockLegalHoldStatus - Specifies whether a legal hold is in effect for this object. This header is only returned if the requester has the s3:GetObjectLegalHold permission. This header is not returned if the specified version of this object has never had a legal hold applied. For more information about S3 Object Lock, see Object Lock.

$sel:cacheControl:HeadObjectResponse', headObjectResponse_cacheControl - Specifies caching behavior along the request/reply chain.

$sel:lastModified:HeadObjectResponse', headObjectResponse_lastModified - Creation date of the object.

$sel:expires:HeadObjectResponse', headObjectResponse_expires - The date and time at which the object is no longer cacheable.

$sel:restore:HeadObjectResponse', headObjectResponse_restore - If the object is an archived object (an object whose storage class is GLACIER), the response includes this header if either the archive restoration is in progress (see RestoreObject or an archive copy is already restored.

If an archive copy is already restored, the header value indicates when Amazon S3 is scheduled to delete the object copy. For example:

x-amz-restore: ongoing-request="false", expiry-date="Fri, 21 Dec 2012 00:00:00 GMT"

If the object restoration is in progress, the header returns the value ongoing-request="true".

For more information about archiving objects, see Transitioning Objects: General Considerations.

$sel:objectLockMode:HeadObjectResponse', headObjectResponse_objectLockMode - The Object Lock mode, if any, that's in effect for this object. This header is only returned if the requester has the s3:GetObjectRetention permission. For more information about S3 Object Lock, see Object Lock.

$sel:objectLockRetainUntilDate:HeadObjectResponse', headObjectResponse_objectLockRetainUntilDate - The date and time when the Object Lock retention period expires. This header is only returned if the requester has the s3:GetObjectRetention permission.

$sel:httpStatus:HeadObjectResponse', headObjectResponse_httpStatus - The response's http status code.

Response Lenses

headObjectResponse_partsCount :: Lens' HeadObjectResponse (Maybe Int)

The count of parts this object has.

headObjectResponse_requestCharged :: Lens' HeadObjectResponse (Maybe RequestCharged)

Undocumented member.

headObjectResponse_eTag :: Lens' HeadObjectResponse (Maybe ETag)

An ETag is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

headObjectResponse_websiteRedirectLocation :: Lens' HeadObjectResponse (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

headObjectResponse_contentType :: Lens' HeadObjectResponse (Maybe Text)

A standard MIME type describing the format of the object data.

headObjectResponse_contentDisposition :: Lens' HeadObjectResponse (Maybe Text)

Specifies presentational information for the object.

headObjectResponse_archiveStatus :: Lens' HeadObjectResponse (Maybe ArchiveStatus)

The archive state of the head object.

headObjectResponse_deleteMarker :: Lens' HeadObjectResponse (Maybe Bool)

Specifies whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

headObjectResponse_contentLanguage :: Lens' HeadObjectResponse (Maybe Text)

The language the content is in.

headObjectResponse_expiration :: Lens' HeadObjectResponse (Maybe Text)

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL encoded.

headObjectResponse_replicationStatus :: Lens' HeadObjectResponse (Maybe ReplicationStatus)

Amazon S3 can return this header if your request involves a bucket that is either a source or a destination in a replication rule.

In replication, you have a source bucket on which you configure replication and destination bucket or buckets where Amazon S3 stores object replicas. When you request an object (GetObject) or object metadata (HeadObject) from these buckets, Amazon S3 will return the x-amz-replication-status header in the response as follows:

If requesting an object from the source bucket — Amazon S3 will return the x-amz-replication-status header if the object in your request is eligible for replication.

For example, suppose that in your replication configuration, you specify object prefix TaxDocs requesting Amazon S3 to replicate objects with key prefix TaxDocs. Any objects you upload with this key name prefix, for example TaxDocs/document1.pdf, are eligible for replication. For any object request with this key name prefix, Amazon S3 will return the x-amz-replication-status header with value PENDING, COMPLETED or FAILED indicating object replication status.

If requesting an object from a destination bucket — Amazon S3 will return the x-amz-replication-status header with value REPLICA if the object in your request is a replica that Amazon S3 created and there is no replica modification replication in progress.

When replicating objects to multiple destination buckets the x-amz-replication-status header acts differently. The header of the source object will only return a value of COMPLETED when replication is successful to all destinations. The header will remain at value PENDING until replication has completed for all destinations. If one or more destinations fails replication the header will return FAILED.

For more information, see Replication.

headObjectResponse_contentLength :: Lens' HeadObjectResponse (Maybe Integer)

Size of the body in bytes.

headObjectResponse_metadata :: Lens' HeadObjectResponse (HashMap Text Text)

A map of metadata to store with the object in S3.

headObjectResponse_sSECustomerKeyMD5 :: Lens' HeadObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

headObjectResponse_sSEKMSKeyId :: Lens' HeadObjectResponse (Maybe Text)

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

headObjectResponse_contentEncoding :: Lens' HeadObjectResponse (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

headObjectResponse_versionId :: Lens' HeadObjectResponse (Maybe ObjectVersionId)

Version of the object.

headObjectResponse_storageClass :: Lens' HeadObjectResponse (Maybe StorageClass)

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

For more information, see Storage Classes.

headObjectResponse_bucketKeyEnabled :: Lens' HeadObjectResponse (Maybe Bool)

Indicates whether the object uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

headObjectResponse_acceptRanges :: Lens' HeadObjectResponse (Maybe Text)

Indicates that a range of bytes was specified.

headObjectResponse_missingMeta :: Lens' HeadObjectResponse (Maybe Int)

This is set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

headObjectResponse_serverSideEncryption :: Lens' HeadObjectResponse (Maybe ServerSideEncryption)

If the object is stored using server-side encryption either with an Amazon Web Services KMS key or an Amazon S3-managed encryption key, the response includes this header with the value of the server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

headObjectResponse_sSECustomerAlgorithm :: Lens' HeadObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

headObjectResponse_objectLockLegalHoldStatus :: Lens' HeadObjectResponse (Maybe ObjectLockLegalHoldStatus)

Specifies whether a legal hold is in effect for this object. This header is only returned if the requester has the s3:GetObjectLegalHold permission. This header is not returned if the specified version of this object has never had a legal hold applied. For more information about S3 Object Lock, see Object Lock.

headObjectResponse_cacheControl :: Lens' HeadObjectResponse (Maybe Text)

Specifies caching behavior along the request/reply chain.

headObjectResponse_lastModified :: Lens' HeadObjectResponse (Maybe UTCTime)

Creation date of the object.

headObjectResponse_expires :: Lens' HeadObjectResponse (Maybe UTCTime)

The date and time at which the object is no longer cacheable.

headObjectResponse_restore :: Lens' HeadObjectResponse (Maybe Text)

If the object is an archived object (an object whose storage class is GLACIER), the response includes this header if either the archive restoration is in progress (see RestoreObject or an archive copy is already restored.

If an archive copy is already restored, the header value indicates when Amazon S3 is scheduled to delete the object copy. For example:

x-amz-restore: ongoing-request="false", expiry-date="Fri, 21 Dec 2012 00:00:00 GMT"

If the object restoration is in progress, the header returns the value ongoing-request="true".

For more information about archiving objects, see Transitioning Objects: General Considerations.

headObjectResponse_objectLockMode :: Lens' HeadObjectResponse (Maybe ObjectLockMode)

The Object Lock mode, if any, that's in effect for this object. This header is only returned if the requester has the s3:GetObjectRetention permission. For more information about S3 Object Lock, see Object Lock.

headObjectResponse_objectLockRetainUntilDate :: Lens' HeadObjectResponse (Maybe UTCTime)

The date and time when the Object Lock retention period expires. This header is only returned if the requester has the s3:GetObjectRetention permission.

headObjectResponse_httpStatus :: Lens' HeadObjectResponse Int

The response's http status code.