Network.AWS.S3.Internal

Description

Documentation

newtype Region

The available AWS regions.

Constructors

Region'

Fields

fromRegion :: Text

Bundled Patterns

pattern Mumbai :: Region

pattern Sydney :: Region

pattern Singapore :: Region

pattern Osaka :: Region

pattern Seoul :: Region

pattern Tokyo :: Region

pattern HongKong :: Region

pattern Ningxia :: Region

pattern Beijing :: Region

pattern CapeTown :: Region

pattern Bahrain :: Region

pattern Stockholm :: Region

pattern Paris :: Region

pattern Milan :: Region

pattern London :: Region

pattern Ireland :: Region

pattern Frankfurt :: Region

pattern SaoPaulo :: Region

pattern Montreal :: Region

pattern GovCloudEast :: Region

pattern GovCloudWest :: Region

pattern Oregon :: Region

pattern NorthCalifornia :: Region

pattern Ohio :: Region

pattern NorthVirginia :: Region

newtype BucketName

Constructors

BucketName Text

newtype ETag

Constructors

ETag ByteString

newtype ObjectVersionId

Constructors

ObjectVersionId Text

Bucket Location

newtype LocationConstraint

Constructors

LocationConstraint

Fields

constraintRegion :: Region

_LocationConstraint :: Iso' LocationConstraint Region

Object Key

type Delimiter = Char

newtype ObjectKey

Constructors

ObjectKey Text

_ObjectKey :: Iso' ObjectKey Text

keyPrefix :: Delimiter -> Traversal' ObjectKey Text

Traverse the prefix of an object key.

The prefix is classified as the entirety of the object key minus the name. A leading prefix in the presence of a name, and no other delimiters is interpreted as a blank prefix.

>>> "/home/jsmith/base.wiki" ^? keyPrefix '/'
Just "/home/jsmith"

>>> "/home/jsmith/" ^? keyPrefix '/'
Just "/home/jsmith"

>>> "/home" ^? keyPrefix '/'
Nothing

>>> "/" ^? keyPrefix '/'
Nothing

keyName :: Delimiter -> Traversal' ObjectKey Text

Traverse the name of an object key.

keyComponents :: Delimiter -> IndexedTraversal' Int ObjectKey Text

Traverse the path components of an object key using the specified delimiter.

Website Endpoints

getWebsiteEndpoint :: Region -> Text

Get the S3 website endpoint for a specific region.

When you configure your bucket as a website, the website is available using this region-specific website endpoint.

See: Amazon Simple Storage Service Website Endpoints.