Network.AWS.S3.ListBucketAnalyticsConfigurations

Description

Lists the analytics configurations for the bucket. You can have up to 1,000 analytics configurations per bucket.

This action supports list pagination and does not return more than 100 configurations at a time. You should always check the IsTruncated element in the response. If there are no more configurations to list, IsTruncated is set to false. If there are more configurations to list, IsTruncated is set to true, and there will be a value in NextContinuationToken. You use the NextContinuationToken value to continue the pagination of the list by passing the value in continuation-token in the request to GET the next page.

To use this operation, you must have permissions to perform the s3:GetAnalyticsConfiguration action. The bucket owner has this permission by default. The bucket owner can grant this permission to others. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to Your Amazon S3 Resources.

For information about Amazon S3 analytics feature, see Amazon S3 Analytics – Storage Class Analysis.

The following operations are related to ListBucketAnalyticsConfigurations:

GetBucketAnalyticsConfiguration

DeleteBucketAnalyticsConfiguration

PutBucketAnalyticsConfiguration

Creating a Request

data ListBucketAnalyticsConfigurations

See: newListBucketAnalyticsConfigurations smart constructor.

Constructors

ListBucketAnalyticsConfigurations'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

continuationToken :: Maybe Text

The ContinuationToken that represents a placeholder from where this request should begin.

bucket :: BucketName

The name of the bucket from which analytics configurations are retrieved.

newListBucketAnalyticsConfigurations

Arguments

:: BucketName
$sel:bucket:ListBucketAnalyticsConfigurations'

-> ListBucketAnalyticsConfigurations

Create a value of ListBucketAnalyticsConfigurations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:ListBucketAnalyticsConfigurations', listBucketAnalyticsConfigurations_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:continuationToken:ListBucketAnalyticsConfigurations', listBucketAnalyticsConfigurations_continuationToken - The ContinuationToken that represents a placeholder from where this request should begin.

$sel:bucket:ListBucketAnalyticsConfigurations', listBucketAnalyticsConfigurations_bucket - The name of the bucket from which analytics configurations are retrieved.

Request Lenses

listBucketAnalyticsConfigurations_expectedBucketOwner :: Lens' ListBucketAnalyticsConfigurations (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

listBucketAnalyticsConfigurations_continuationToken :: Lens' ListBucketAnalyticsConfigurations (Maybe Text)

The ContinuationToken that represents a placeholder from where this request should begin.

listBucketAnalyticsConfigurations_bucket :: Lens' ListBucketAnalyticsConfigurations BucketName

The name of the bucket from which analytics configurations are retrieved.

Destructuring the Response

data ListBucketAnalyticsConfigurationsResponse

See: newListBucketAnalyticsConfigurationsResponse smart constructor.

Constructors

ListBucketAnalyticsConfigurationsResponse'

Fields

analyticsConfigurationList :: Maybe [AnalyticsConfiguration]

The list of analytics configurations for a bucket.

isTruncated :: Maybe Bool

Indicates whether the returned list of analytics configurations is complete. A value of true indicates that the list is not complete and the NextContinuationToken will be provided for a subsequent request.

nextContinuationToken :: Maybe Text

NextContinuationToken is sent when isTruncated is true, which indicates that there are more analytics configurations to list. The next request must include this NextContinuationToken. The token is obfuscated and is not a usable value.

continuationToken :: Maybe Text

The marker that is used as a starting point for this analytics configuration list response. This value is present if it was sent in the request.

httpStatus :: Int

The response's http status code.

newListBucketAnalyticsConfigurationsResponse

Arguments

:: Int
$sel:httpStatus:ListBucketAnalyticsConfigurationsResponse'

-> ListBucketAnalyticsConfigurationsResponse

Create a value of ListBucketAnalyticsConfigurationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:analyticsConfigurationList:ListBucketAnalyticsConfigurationsResponse', listBucketAnalyticsConfigurationsResponse_analyticsConfigurationList - The list of analytics configurations for a bucket.

$sel:isTruncated:ListBucketAnalyticsConfigurationsResponse', listBucketAnalyticsConfigurationsResponse_isTruncated - Indicates whether the returned list of analytics configurations is complete. A value of true indicates that the list is not complete and the NextContinuationToken will be provided for a subsequent request.

$sel:nextContinuationToken:ListBucketAnalyticsConfigurationsResponse', listBucketAnalyticsConfigurationsResponse_nextContinuationToken - NextContinuationToken is sent when isTruncated is true, which indicates that there are more analytics configurations to list. The next request must include this NextContinuationToken. The token is obfuscated and is not a usable value.

$sel:continuationToken:ListBucketAnalyticsConfigurations', listBucketAnalyticsConfigurationsResponse_continuationToken - The marker that is used as a starting point for this analytics configuration list response. This value is present if it was sent in the request.

$sel:httpStatus:ListBucketAnalyticsConfigurationsResponse', listBucketAnalyticsConfigurationsResponse_httpStatus - The response's http status code.

Response Lenses

listBucketAnalyticsConfigurationsResponse_analyticsConfigurationList :: Lens' ListBucketAnalyticsConfigurationsResponse (Maybe [AnalyticsConfiguration])

The list of analytics configurations for a bucket.

listBucketAnalyticsConfigurationsResponse_isTruncated :: Lens' ListBucketAnalyticsConfigurationsResponse (Maybe Bool)

Indicates whether the returned list of analytics configurations is complete. A value of true indicates that the list is not complete and the NextContinuationToken will be provided for a subsequent request.

listBucketAnalyticsConfigurationsResponse_nextContinuationToken :: Lens' ListBucketAnalyticsConfigurationsResponse (Maybe Text)

NextContinuationToken is sent when isTruncated is true, which indicates that there are more analytics configurations to list. The next request must include this NextContinuationToken. The token is obfuscated and is not a usable value.

listBucketAnalyticsConfigurationsResponse_continuationToken :: Lens' ListBucketAnalyticsConfigurationsResponse (Maybe Text)

The marker that is used as a starting point for this analytics configuration list response. This value is present if it was sent in the request.

listBucketAnalyticsConfigurationsResponse_httpStatus :: Lens' ListBucketAnalyticsConfigurationsResponse Int

The response's http status code.