Network.AWS.S3.ListBucketInventoryConfigurations

Description

Returns a list of inventory configurations for the bucket. You can have up to 1,000 analytics configurations per bucket.

This action supports list pagination and does not return more than 100 configurations at a time. Always check the IsTruncated element in the response. If there are no more configurations to list, IsTruncated is set to false. If there are more configurations to list, IsTruncated is set to true, and there is a value in NextContinuationToken. You use the NextContinuationToken value to continue the pagination of the list by passing the value in continuation-token in the request to GET the next page.

To use this operation, you must have permissions to perform the s3:GetInventoryConfiguration action. The bucket owner has this permission by default. The bucket owner can grant this permission to others. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to Your Amazon S3 Resources.

For information about the Amazon S3 inventory feature, see Amazon S3 Inventory

The following operations are related to ListBucketInventoryConfigurations:

GetBucketInventoryConfiguration

DeleteBucketInventoryConfiguration

PutBucketInventoryConfiguration

Creating a Request

data ListBucketInventoryConfigurations

See: newListBucketInventoryConfigurations smart constructor.

Constructors

ListBucketInventoryConfigurations'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

continuationToken :: Maybe Text

The marker used to continue an inventory configuration listing that has been truncated. Use the NextContinuationToken from a previously truncated list response to continue the listing. The continuation token is an opaque value that Amazon S3 understands.

bucket :: BucketName

The name of the bucket containing the inventory configurations to retrieve.

newListBucketInventoryConfigurations

Arguments

:: BucketName
$sel:bucket:ListBucketInventoryConfigurations'

-> ListBucketInventoryConfigurations

Create a value of ListBucketInventoryConfigurations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:ListBucketInventoryConfigurations', listBucketInventoryConfigurations_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:continuationToken:ListBucketInventoryConfigurations', listBucketInventoryConfigurations_continuationToken - The marker used to continue an inventory configuration listing that has been truncated. Use the NextContinuationToken from a previously truncated list response to continue the listing. The continuation token is an opaque value that Amazon S3 understands.

$sel:bucket:ListBucketInventoryConfigurations', listBucketInventoryConfigurations_bucket - The name of the bucket containing the inventory configurations to retrieve.

Request Lenses

listBucketInventoryConfigurations_expectedBucketOwner :: Lens' ListBucketInventoryConfigurations (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

listBucketInventoryConfigurations_continuationToken :: Lens' ListBucketInventoryConfigurations (Maybe Text)

The marker used to continue an inventory configuration listing that has been truncated. Use the NextContinuationToken from a previously truncated list response to continue the listing. The continuation token is an opaque value that Amazon S3 understands.

listBucketInventoryConfigurations_bucket :: Lens' ListBucketInventoryConfigurations BucketName

The name of the bucket containing the inventory configurations to retrieve.

Destructuring the Response

data ListBucketInventoryConfigurationsResponse

See: newListBucketInventoryConfigurationsResponse smart constructor.

Constructors

ListBucketInventoryConfigurationsResponse'

Fields

inventoryConfigurationList :: Maybe [InventoryConfiguration]

The list of inventory configurations for a bucket.

isTruncated :: Maybe Bool

Tells whether the returned list of inventory configurations is complete. A value of true indicates that the list is not complete and the NextContinuationToken is provided for a subsequent request.

nextContinuationToken :: Maybe Text

The marker used to continue this inventory configuration listing. Use the NextContinuationToken from this response to continue the listing in a subsequent request. The continuation token is an opaque value that Amazon S3 understands.

continuationToken :: Maybe Text

If sent in the request, the marker that is used as a starting point for this inventory configuration list response.

httpStatus :: Int

The response's http status code.

newListBucketInventoryConfigurationsResponse

Arguments

:: Int
$sel:httpStatus:ListBucketInventoryConfigurationsResponse'

-> ListBucketInventoryConfigurationsResponse

Create a value of ListBucketInventoryConfigurationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inventoryConfigurationList:ListBucketInventoryConfigurationsResponse', listBucketInventoryConfigurationsResponse_inventoryConfigurationList - The list of inventory configurations for a bucket.

$sel:isTruncated:ListBucketInventoryConfigurationsResponse', listBucketInventoryConfigurationsResponse_isTruncated - Tells whether the returned list of inventory configurations is complete. A value of true indicates that the list is not complete and the NextContinuationToken is provided for a subsequent request.

$sel:nextContinuationToken:ListBucketInventoryConfigurationsResponse', listBucketInventoryConfigurationsResponse_nextContinuationToken - The marker used to continue this inventory configuration listing. Use the NextContinuationToken from this response to continue the listing in a subsequent request. The continuation token is an opaque value that Amazon S3 understands.

$sel:continuationToken:ListBucketInventoryConfigurations', listBucketInventoryConfigurationsResponse_continuationToken - If sent in the request, the marker that is used as a starting point for this inventory configuration list response.

$sel:httpStatus:ListBucketInventoryConfigurationsResponse', listBucketInventoryConfigurationsResponse_httpStatus - The response's http status code.

Response Lenses

listBucketInventoryConfigurationsResponse_inventoryConfigurationList :: Lens' ListBucketInventoryConfigurationsResponse (Maybe [InventoryConfiguration])

The list of inventory configurations for a bucket.

listBucketInventoryConfigurationsResponse_isTruncated :: Lens' ListBucketInventoryConfigurationsResponse (Maybe Bool)

Tells whether the returned list of inventory configurations is complete. A value of true indicates that the list is not complete and the NextContinuationToken is provided for a subsequent request.

listBucketInventoryConfigurationsResponse_nextContinuationToken :: Lens' ListBucketInventoryConfigurationsResponse (Maybe Text)

The marker used to continue this inventory configuration listing. Use the NextContinuationToken from this response to continue the listing in a subsequent request. The continuation token is an opaque value that Amazon S3 understands.

listBucketInventoryConfigurationsResponse_continuationToken :: Lens' ListBucketInventoryConfigurationsResponse (Maybe Text)

If sent in the request, the marker that is used as a starting point for this inventory configuration list response.

listBucketInventoryConfigurationsResponse_httpStatus :: Lens' ListBucketInventoryConfigurationsResponse Int

The response's http status code.