Network.AWS.S3.ListBucketMetricsConfigurations

Description

Lists the metrics configurations for the bucket. The metrics configurations are only for the request metrics of the bucket and do not provide information on daily storage metrics. You can have up to 1,000 configurations per bucket.

This action supports list pagination and does not return more than 100 configurations at a time. Always check the IsTruncated element in the response. If there are no more configurations to list, IsTruncated is set to false. If there are more configurations to list, IsTruncated is set to true, and there is a value in NextContinuationToken. You use the NextContinuationToken value to continue the pagination of the list by passing the value in continuation-token in the request to GET the next page.

To use this operation, you must have permissions to perform the s3:GetMetricsConfiguration action. The bucket owner has this permission by default. The bucket owner can grant this permission to others. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to Your Amazon S3 Resources.

For more information about metrics configurations and CloudWatch request metrics, see Monitoring Metrics with Amazon CloudWatch.

The following operations are related to ListBucketMetricsConfigurations:

PutBucketMetricsConfiguration

GetBucketMetricsConfiguration

DeleteBucketMetricsConfiguration

Creating a Request

data ListBucketMetricsConfigurations

See: newListBucketMetricsConfigurations smart constructor.

Constructors

ListBucketMetricsConfigurations'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

continuationToken :: Maybe Text

The marker that is used to continue a metrics configuration listing that has been truncated. Use the NextContinuationToken from a previously truncated list response to continue the listing. The continuation token is an opaque value that Amazon S3 understands.

bucket :: BucketName

The name of the bucket containing the metrics configurations to retrieve.

newListBucketMetricsConfigurations

Arguments

:: BucketName
$sel:bucket:ListBucketMetricsConfigurations'

-> ListBucketMetricsConfigurations

Create a value of ListBucketMetricsConfigurations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:ListBucketMetricsConfigurations', listBucketMetricsConfigurations_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:continuationToken:ListBucketMetricsConfigurations', listBucketMetricsConfigurations_continuationToken - The marker that is used to continue a metrics configuration listing that has been truncated. Use the NextContinuationToken from a previously truncated list response to continue the listing. The continuation token is an opaque value that Amazon S3 understands.

$sel:bucket:ListBucketMetricsConfigurations', listBucketMetricsConfigurations_bucket - The name of the bucket containing the metrics configurations to retrieve.

Request Lenses

listBucketMetricsConfigurations_expectedBucketOwner :: Lens' ListBucketMetricsConfigurations (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

listBucketMetricsConfigurations_continuationToken :: Lens' ListBucketMetricsConfigurations (Maybe Text)

The marker that is used to continue a metrics configuration listing that has been truncated. Use the NextContinuationToken from a previously truncated list response to continue the listing. The continuation token is an opaque value that Amazon S3 understands.

listBucketMetricsConfigurations_bucket :: Lens' ListBucketMetricsConfigurations BucketName

The name of the bucket containing the metrics configurations to retrieve.

Destructuring the Response

data ListBucketMetricsConfigurationsResponse

See: newListBucketMetricsConfigurationsResponse smart constructor.

Constructors

ListBucketMetricsConfigurationsResponse'

Fields

isTruncated :: Maybe Bool

Indicates whether the returned list of metrics configurations is complete. A value of true indicates that the list is not complete and the NextContinuationToken will be provided for a subsequent request.

metricsConfigurationList :: Maybe [MetricsConfiguration]

The list of metrics configurations for a bucket.

nextContinuationToken :: Maybe Text

The marker used to continue a metrics configuration listing that has been truncated. Use the NextContinuationToken from a previously truncated list response to continue the listing. The continuation token is an opaque value that Amazon S3 understands.

continuationToken :: Maybe Text

The marker that is used as a starting point for this metrics configuration list response. This value is present if it was sent in the request.

httpStatus :: Int

The response's http status code.

newListBucketMetricsConfigurationsResponse

Arguments

:: Int
$sel:httpStatus:ListBucketMetricsConfigurationsResponse'

-> ListBucketMetricsConfigurationsResponse

Create a value of ListBucketMetricsConfigurationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isTruncated:ListBucketMetricsConfigurationsResponse', listBucketMetricsConfigurationsResponse_isTruncated - Indicates whether the returned list of metrics configurations is complete. A value of true indicates that the list is not complete and the NextContinuationToken will be provided for a subsequent request.

$sel:metricsConfigurationList:ListBucketMetricsConfigurationsResponse', listBucketMetricsConfigurationsResponse_metricsConfigurationList - The list of metrics configurations for a bucket.

$sel:nextContinuationToken:ListBucketMetricsConfigurationsResponse', listBucketMetricsConfigurationsResponse_nextContinuationToken - The marker used to continue a metrics configuration listing that has been truncated. Use the NextContinuationToken from a previously truncated list response to continue the listing. The continuation token is an opaque value that Amazon S3 understands.

$sel:continuationToken:ListBucketMetricsConfigurations', listBucketMetricsConfigurationsResponse_continuationToken - The marker that is used as a starting point for this metrics configuration list response. This value is present if it was sent in the request.

$sel:httpStatus:ListBucketMetricsConfigurationsResponse', listBucketMetricsConfigurationsResponse_httpStatus - The response's http status code.

Response Lenses

listBucketMetricsConfigurationsResponse_isTruncated :: Lens' ListBucketMetricsConfigurationsResponse (Maybe Bool)

Indicates whether the returned list of metrics configurations is complete. A value of true indicates that the list is not complete and the NextContinuationToken will be provided for a subsequent request.

listBucketMetricsConfigurationsResponse_metricsConfigurationList :: Lens' ListBucketMetricsConfigurationsResponse (Maybe [MetricsConfiguration])

The list of metrics configurations for a bucket.

listBucketMetricsConfigurationsResponse_nextContinuationToken :: Lens' ListBucketMetricsConfigurationsResponse (Maybe Text)

The marker used to continue a metrics configuration listing that has been truncated. Use the NextContinuationToken from a previously truncated list response to continue the listing. The continuation token is an opaque value that Amazon S3 understands.

listBucketMetricsConfigurationsResponse_continuationToken :: Lens' ListBucketMetricsConfigurationsResponse (Maybe Text)

The marker that is used as a starting point for this metrics configuration list response. This value is present if it was sent in the request.

listBucketMetricsConfigurationsResponse_httpStatus :: Lens' ListBucketMetricsConfigurationsResponse Int

The response's http status code.