Network.AWS.S3.ListMultipartUploads

Description

This action lists in-progress multipart uploads. An in-progress multipart upload is a multipart upload that has been initiated using the Initiate Multipart Upload request, but has not yet been completed or aborted.

This action returns at most 1,000 multipart uploads in the response. 1,000 multipart uploads is the maximum number of uploads a response can include, which is also the default value. You can further limit the number of uploads in a response by specifying the max-uploads parameter in the response. If additional multipart uploads satisfy the list criteria, the response will contain an IsTruncated element with the value true. To list the additional multipart uploads, use the key-marker and upload-id-marker request parameters.

In the response, the uploads are sorted by key. If your application has initiated more than one multipart upload using the same object key, then uploads in the response are first sorted by key. Additionally, uploads are sorted in ascending order within each key by the upload initiation time.

For more information on multipart uploads, see Uploading Objects Using Multipart Upload.

For information on permissions required to use the multipart upload API, see Multipart Upload and Permissions.

The following operations are related to ListMultipartUploads:

CreateMultipartUpload

UploadPart

CompleteMultipartUpload

ListParts

AbortMultipartUpload

This operation returns paginated results.

Creating a Request

data ListMultipartUploads

See: newListMultipartUploads smart constructor.

Constructors

ListMultipartUploads'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

delimiter :: Maybe Delimiter

Character you use to group keys.

All keys that contain the same string between the prefix, if specified, and the first occurrence of the delimiter after the prefix are grouped under a single result element, CommonPrefixes. If you don't specify the prefix parameter, then the substring starts at the beginning of the key. The keys that are grouped under CommonPrefixes result element are not returned elsewhere in the response.

encodingType :: Maybe EncodingType

prefix :: Maybe Text

Lists in-progress uploads only for those keys that begin with the specified prefix. You can use prefixes to separate a bucket into different grouping of keys. (You can think of using prefix to make groups in the same way you'd use a folder in a file system.)

keyMarker :: Maybe Text

Together with upload-id-marker, this parameter specifies the multipart upload after which listing should begin.

If upload-id-marker is not specified, only the keys lexicographically greater than the specified key-marker will be included in the list.

If upload-id-marker is specified, any multipart uploads for a key equal to the key-marker might also be included, provided those multipart uploads have upload IDs lexicographically greater than the specified upload-id-marker.

uploadIdMarker :: Maybe Text

Together with key-marker, specifies the multipart upload after which listing should begin. If key-marker is not specified, the upload-id-marker parameter is ignored. Otherwise, any multipart uploads for a key equal to the key-marker might be included in the list only if they have an upload ID lexicographically greater than the specified upload-id-marker.

maxUploads :: Maybe Int

Sets the maximum number of multipart uploads, from 1 to 1,000, to return in the response body. 1,000 is the maximum number of uploads that can be returned in a response.

bucket :: BucketName

The name of the bucket to which the multipart upload was initiated.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

newListMultipartUploads

Arguments

:: BucketName
$sel:bucket:ListMultipartUploads'

-> ListMultipartUploads

Create a value of ListMultipartUploads with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:ListMultipartUploads', listMultipartUploads_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:delimiter:ListMultipartUploads', listMultipartUploads_delimiter - Character you use to group keys.

All keys that contain the same string between the prefix, if specified, and the first occurrence of the delimiter after the prefix are grouped under a single result element, CommonPrefixes. If you don't specify the prefix parameter, then the substring starts at the beginning of the key. The keys that are grouped under CommonPrefixes result element are not returned elsewhere in the response.

$sel:encodingType:ListMultipartUploads', listMultipartUploads_encodingType - Undocumented member.

$sel:prefix:ListMultipartUploads', listMultipartUploads_prefix - Lists in-progress uploads only for those keys that begin with the specified prefix. You can use prefixes to separate a bucket into different grouping of keys. (You can think of using prefix to make groups in the same way you'd use a folder in a file system.)

$sel:keyMarker:ListMultipartUploads', listMultipartUploads_keyMarker - Together with upload-id-marker, this parameter specifies the multipart upload after which listing should begin.

If upload-id-marker is not specified, only the keys lexicographically greater than the specified key-marker will be included in the list.

If upload-id-marker is specified, any multipart uploads for a key equal to the key-marker might also be included, provided those multipart uploads have upload IDs lexicographically greater than the specified upload-id-marker.

$sel:uploadIdMarker:ListMultipartUploads', listMultipartUploads_uploadIdMarker - Together with key-marker, specifies the multipart upload after which listing should begin. If key-marker is not specified, the upload-id-marker parameter is ignored. Otherwise, any multipart uploads for a key equal to the key-marker might be included in the list only if they have an upload ID lexicographically greater than the specified upload-id-marker.

$sel:maxUploads:ListMultipartUploads', listMultipartUploads_maxUploads - Sets the maximum number of multipart uploads, from 1 to 1,000, to return in the response body. 1,000 is the maximum number of uploads that can be returned in a response.

$sel:bucket:ListMultipartUploads', listMultipartUploads_bucket - The name of the bucket to which the multipart upload was initiated.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

Request Lenses

listMultipartUploads_expectedBucketOwner :: Lens' ListMultipartUploads (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

listMultipartUploads_delimiter :: Lens' ListMultipartUploads (Maybe Delimiter)

Character you use to group keys.

All keys that contain the same string between the prefix, if specified, and the first occurrence of the delimiter after the prefix are grouped under a single result element, CommonPrefixes. If you don't specify the prefix parameter, then the substring starts at the beginning of the key. The keys that are grouped under CommonPrefixes result element are not returned elsewhere in the response.

listMultipartUploads_encodingType :: Lens' ListMultipartUploads (Maybe EncodingType)

Undocumented member.

listMultipartUploads_prefix :: Lens' ListMultipartUploads (Maybe Text)

Lists in-progress uploads only for those keys that begin with the specified prefix. You can use prefixes to separate a bucket into different grouping of keys. (You can think of using prefix to make groups in the same way you'd use a folder in a file system.)

listMultipartUploads_keyMarker :: Lens' ListMultipartUploads (Maybe Text)

Together with upload-id-marker, this parameter specifies the multipart upload after which listing should begin.

If upload-id-marker is not specified, only the keys lexicographically greater than the specified key-marker will be included in the list.

If upload-id-marker is specified, any multipart uploads for a key equal to the key-marker might also be included, provided those multipart uploads have upload IDs lexicographically greater than the specified upload-id-marker.

listMultipartUploads_uploadIdMarker :: Lens' ListMultipartUploads (Maybe Text)

Together with key-marker, specifies the multipart upload after which listing should begin. If key-marker is not specified, the upload-id-marker parameter is ignored. Otherwise, any multipart uploads for a key equal to the key-marker might be included in the list only if they have an upload ID lexicographically greater than the specified upload-id-marker.

listMultipartUploads_maxUploads :: Lens' ListMultipartUploads (Maybe Int)

Sets the maximum number of multipart uploads, from 1 to 1,000, to return in the response body. 1,000 is the maximum number of uploads that can be returned in a response.

listMultipartUploads_bucket :: Lens' ListMultipartUploads BucketName

The name of the bucket to which the multipart upload was initiated.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

Destructuring the Response

data ListMultipartUploadsResponse

See: newListMultipartUploadsResponse smart constructor.

Constructors

ListMultipartUploadsResponse'

Fields

commonPrefixes :: Maybe [CommonPrefix]

If you specify a delimiter in the request, then the result returns each distinct key prefix containing the delimiter in a CommonPrefixes element. The distinct key prefixes are returned in the Prefix child element.

delimiter :: Maybe Delimiter

Contains the delimiter you specified in the request. If you don't specify a delimiter in your request, this element is absent from the response.

encodingType :: Maybe EncodingType

Encoding type used by Amazon S3 to encode object keys in the response.

If you specify encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

Delimiter, KeyMarker, Prefix, NextKeyMarker, Key.

prefix :: Maybe Text

When a prefix is provided in the request, this field contains the specified prefix. The result contains only keys starting with the specified prefix.

isTruncated :: Maybe Bool

Indicates whether the returned list of multipart uploads is truncated. A value of true indicates that the list was truncated. The list can be truncated if the number of multipart uploads exceeds the limit allowed or specified by max uploads.

uploads :: Maybe [MultipartUpload]

Container for elements related to a particular multipart upload. A response can contain zero or more Upload elements.

keyMarker :: Maybe Text

The key at or after which the listing began.

nextKeyMarker :: Maybe Text

When a list is truncated, this element specifies the value that should be used for the key-marker request parameter in a subsequent request.

uploadIdMarker :: Maybe Text

Upload ID after which listing began.

maxUploads :: Maybe Int

Maximum number of multipart uploads that could have been included in the response.

bucket :: Maybe BucketName

The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

nextUploadIdMarker :: Maybe Text

When a list is truncated, this element specifies the value that should be used for the upload-id-marker request parameter in a subsequent request.

httpStatus :: Int

The response's http status code.

newListMultipartUploadsResponse

Arguments

:: Int
$sel:httpStatus:ListMultipartUploadsResponse'

-> ListMultipartUploadsResponse

Create a value of ListMultipartUploadsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:commonPrefixes:ListMultipartUploadsResponse', listMultipartUploadsResponse_commonPrefixes - If you specify a delimiter in the request, then the result returns each distinct key prefix containing the delimiter in a CommonPrefixes element. The distinct key prefixes are returned in the Prefix child element.

$sel:delimiter:ListMultipartUploads', listMultipartUploadsResponse_delimiter - Contains the delimiter you specified in the request. If you don't specify a delimiter in your request, this element is absent from the response.

$sel:encodingType:ListMultipartUploads', listMultipartUploadsResponse_encodingType - Encoding type used by Amazon S3 to encode object keys in the response.

If you specify encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

Delimiter, KeyMarker, Prefix, NextKeyMarker, Key.

$sel:prefix:ListMultipartUploads', listMultipartUploadsResponse_prefix - When a prefix is provided in the request, this field contains the specified prefix. The result contains only keys starting with the specified prefix.

$sel:isTruncated:ListMultipartUploadsResponse', listMultipartUploadsResponse_isTruncated - Indicates whether the returned list of multipart uploads is truncated. A value of true indicates that the list was truncated. The list can be truncated if the number of multipart uploads exceeds the limit allowed or specified by max uploads.

$sel:uploads:ListMultipartUploadsResponse', listMultipartUploadsResponse_uploads - Container for elements related to a particular multipart upload. A response can contain zero or more Upload elements.

$sel:keyMarker:ListMultipartUploads', listMultipartUploadsResponse_keyMarker - The key at or after which the listing began.

$sel:nextKeyMarker:ListMultipartUploadsResponse', listMultipartUploadsResponse_nextKeyMarker - When a list is truncated, this element specifies the value that should be used for the key-marker request parameter in a subsequent request.

$sel:uploadIdMarker:ListMultipartUploads', listMultipartUploadsResponse_uploadIdMarker - Upload ID after which listing began.

$sel:maxUploads:ListMultipartUploads', listMultipartUploadsResponse_maxUploads - Maximum number of multipart uploads that could have been included in the response.

$sel:bucket:ListMultipartUploads', listMultipartUploadsResponse_bucket - The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

$sel:nextUploadIdMarker:ListMultipartUploadsResponse', listMultipartUploadsResponse_nextUploadIdMarker - When a list is truncated, this element specifies the value that should be used for the upload-id-marker request parameter in a subsequent request.

$sel:httpStatus:ListMultipartUploadsResponse', listMultipartUploadsResponse_httpStatus - The response's http status code.

Response Lenses

listMultipartUploadsResponse_commonPrefixes :: Lens' ListMultipartUploadsResponse (Maybe [CommonPrefix])

If you specify a delimiter in the request, then the result returns each distinct key prefix containing the delimiter in a CommonPrefixes element. The distinct key prefixes are returned in the Prefix child element.

listMultipartUploadsResponse_delimiter :: Lens' ListMultipartUploadsResponse (Maybe Delimiter)

Contains the delimiter you specified in the request. If you don't specify a delimiter in your request, this element is absent from the response.

listMultipartUploadsResponse_encodingType :: Lens' ListMultipartUploadsResponse (Maybe EncodingType)

Encoding type used by Amazon S3 to encode object keys in the response.

If you specify encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

Delimiter, KeyMarker, Prefix, NextKeyMarker, Key.

listMultipartUploadsResponse_prefix :: Lens' ListMultipartUploadsResponse (Maybe Text)

When a prefix is provided in the request, this field contains the specified prefix. The result contains only keys starting with the specified prefix.

listMultipartUploadsResponse_isTruncated :: Lens' ListMultipartUploadsResponse (Maybe Bool)

Indicates whether the returned list of multipart uploads is truncated. A value of true indicates that the list was truncated. The list can be truncated if the number of multipart uploads exceeds the limit allowed or specified by max uploads.

listMultipartUploadsResponse_uploads :: Lens' ListMultipartUploadsResponse (Maybe [MultipartUpload])

Container for elements related to a particular multipart upload. A response can contain zero or more Upload elements.

listMultipartUploadsResponse_keyMarker :: Lens' ListMultipartUploadsResponse (Maybe Text)

The key at or after which the listing began.

listMultipartUploadsResponse_nextKeyMarker :: Lens' ListMultipartUploadsResponse (Maybe Text)

When a list is truncated, this element specifies the value that should be used for the key-marker request parameter in a subsequent request.

listMultipartUploadsResponse_uploadIdMarker :: Lens' ListMultipartUploadsResponse (Maybe Text)

Upload ID after which listing began.

listMultipartUploadsResponse_maxUploads :: Lens' ListMultipartUploadsResponse (Maybe Int)

Maximum number of multipart uploads that could have been included in the response.

listMultipartUploadsResponse_bucket :: Lens' ListMultipartUploadsResponse (Maybe BucketName)

The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

listMultipartUploadsResponse_nextUploadIdMarker :: Lens' ListMultipartUploadsResponse (Maybe Text)

When a list is truncated, this element specifies the value that should be used for the upload-id-marker request parameter in a subsequent request.

listMultipartUploadsResponse_httpStatus :: Lens' ListMultipartUploadsResponse Int

The response's http status code.