Network.AWS.S3.ListObjectsV

Description

Returns some or all (up to 1,000) of the objects in a bucket with each request. You can use the request parameters as selection criteria to return a subset of the objects in a bucket. A 200 OK response can contain valid or invalid XML. Make sure to design your application to parse the contents of the response and handle it appropriately. Objects are returned sorted in an ascending order of the respective key names in the list. For more information about listing objects, see Listing object keys programmatically

To use this operation, you must have READ access to the bucket.

To use this action in an Identity and Access Management (IAM) policy, you must have permissions to perform the s3:ListBucket action. The bucket owner has this permission by default and can grant this permission to others. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to Your Amazon S3 Resources.

This section describes the latest revision of this action. We recommend that you use this revised API for application development. For backward compatibility, Amazon S3 continues to support the prior version of this API, ListObjects.

To get a list of your buckets, see ListBuckets.

The following operations are related to ListObjectsV2:

GetObject

PutObject

CreateBucket

Creating a Request

data ListObjectsV

See: newListObjectsV smart constructor.

Constructors

ListObjectsV'

Fields

startAfter :: Maybe Text

StartAfter is where you want Amazon S3 to start listing from. Amazon S3 starts listing after this specified key. StartAfter can be any key in the bucket.

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

delimiter :: Maybe Delimiter

A delimiter is a character you use to group keys.

encodingType :: Maybe EncodingType

Encoding type used by Amazon S3 to encode object keys in the response.

prefix :: Maybe Text

Limits the response to keys that begin with the specified prefix.

maxKeys :: Maybe Int

Sets the maximum number of keys returned in the response. By default the action returns up to 1,000 key names. The response might contain fewer keys but will never contain more.

requestPayer :: Maybe RequestPayer

Confirms that the requester knows that she or he will be charged for the list objects request in V2 style. Bucket owners need not specify this parameter in their requests.

fetchOwner :: Maybe Bool

The owner field is not present in listV2 by default, if you want to return owner field with each key in the result then set the fetch owner field to true.

continuationToken :: Maybe Text

ContinuationToken indicates Amazon S3 that the list is being continued on this bucket with a token. ContinuationToken is obfuscated and is not a real key.

bucket :: BucketName

Bucket name to list.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

newListObjectsV

Arguments

:: BucketName
$sel:bucket:ListObjectsV'

-> ListObjectsV

Create a value of ListObjectsV with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startAfter:ListObjectsV', listObjectsV_startAfter - StartAfter is where you want Amazon S3 to start listing from. Amazon S3 starts listing after this specified key. StartAfter can be any key in the bucket.

$sel:expectedBucketOwner:ListObjectsV', listObjectsV_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:delimiter:ListObjectsV', listObjectsV_delimiter - A delimiter is a character you use to group keys.

$sel:encodingType:ListObjectsV', listObjectsV_encodingType - Encoding type used by Amazon S3 to encode object keys in the response.

$sel:prefix:ListObjectsV', listObjectsV_prefix - Limits the response to keys that begin with the specified prefix.

$sel:maxKeys:ListObjectsV', listObjectsV_maxKeys - Sets the maximum number of keys returned in the response. By default the action returns up to 1,000 key names. The response might contain fewer keys but will never contain more.

$sel:requestPayer:ListObjectsV', listObjectsV_requestPayer - Confirms that the requester knows that she or he will be charged for the list objects request in V2 style. Bucket owners need not specify this parameter in their requests.

$sel:fetchOwner:ListObjectsV', listObjectsV_fetchOwner - The owner field is not present in listV2 by default, if you want to return owner field with each key in the result then set the fetch owner field to true.

$sel:continuationToken:ListObjectsV', listObjectsV_continuationToken - ContinuationToken indicates Amazon S3 that the list is being continued on this bucket with a token. ContinuationToken is obfuscated and is not a real key.

$sel:bucket:ListObjectsV', listObjectsV_bucket - Bucket name to list.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

Request Lenses

listObjectsV_startAfter :: Lens' ListObjectsV (Maybe Text)

StartAfter is where you want Amazon S3 to start listing from. Amazon S3 starts listing after this specified key. StartAfter can be any key in the bucket.

listObjectsV_expectedBucketOwner :: Lens' ListObjectsV (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

listObjectsV_delimiter :: Lens' ListObjectsV (Maybe Delimiter)

A delimiter is a character you use to group keys.

listObjectsV_encodingType :: Lens' ListObjectsV (Maybe EncodingType)

Encoding type used by Amazon S3 to encode object keys in the response.

listObjectsV_prefix :: Lens' ListObjectsV (Maybe Text)

Limits the response to keys that begin with the specified prefix.

listObjectsV_maxKeys :: Lens' ListObjectsV (Maybe Int)

Sets the maximum number of keys returned in the response. By default the action returns up to 1,000 key names. The response might contain fewer keys but will never contain more.

listObjectsV_requestPayer :: Lens' ListObjectsV (Maybe RequestPayer)

Confirms that the requester knows that she or he will be charged for the list objects request in V2 style. Bucket owners need not specify this parameter in their requests.

listObjectsV_fetchOwner :: Lens' ListObjectsV (Maybe Bool)

The owner field is not present in listV2 by default, if you want to return owner field with each key in the result then set the fetch owner field to true.

listObjectsV_continuationToken :: Lens' ListObjectsV (Maybe Text)

ContinuationToken indicates Amazon S3 that the list is being continued on this bucket with a token. ContinuationToken is obfuscated and is not a real key.

listObjectsV_bucket :: Lens' ListObjectsV BucketName

Bucket name to list.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

Destructuring the Response

data ListObjectsVResponse

See: newListObjectsVResponse smart constructor.

Constructors

ListObjectsVResponse'

Fields

keyCount :: Maybe Int

KeyCount is the number of keys returned with this request. KeyCount will always be less than or equals to MaxKeys field. Say you ask for 50 keys, your result will include less than equals 50 keys

startAfter :: Maybe Text

If StartAfter was sent with the request, it is included in the response.

commonPrefixes :: Maybe [CommonPrefix]

All of the keys (up to 1,000) rolled up into a common prefix count as a single return when calculating the number of returns.

A response can contain CommonPrefixes only if you specify a delimiter.

CommonPrefixes contains all (if there are any) keys between Prefix and the next occurrence of the string specified by a delimiter.

CommonPrefixes lists keys that act like subdirectories in the directory specified by Prefix.

For example, if the prefix is notes/ and the delimiter is a slash (/) as in notes/summer/july, the common prefix is notes/summer/. All of the keys that roll up into a common prefix count as a single return when calculating the number of returns.

delimiter :: Maybe Delimiter

Causes keys that contain the same string between the prefix and the first occurrence of the delimiter to be rolled up into a single result element in the CommonPrefixes collection. These rolled-up keys are not returned elsewhere in the response. Each rolled-up result counts as only one return against the MaxKeys value.

encodingType :: Maybe EncodingType

Encoding type used by Amazon S3 to encode object key names in the XML response.

If you specify the encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

Delimiter, Prefix, Key, and StartAfter.

prefix :: Maybe Text

Keys that begin with the indicated prefix.

maxKeys :: Maybe Int

Sets the maximum number of keys returned in the response. By default the action returns up to 1,000 key names. The response might contain fewer keys but will never contain more.

isTruncated :: Maybe Bool

Set to false if all of the results were returned. Set to true if more keys are available to return. If the number of results exceeds that specified by MaxKeys, all of the results might not be returned.

contents :: Maybe [Object]

Metadata about each object returned.

name :: Maybe BucketName

The bucket name.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

nextContinuationToken :: Maybe Text

NextContinuationToken is sent when isTruncated is true, which means there are more keys in the bucket that can be listed. The next list requests to Amazon S3 can be continued with this NextContinuationToken. NextContinuationToken is obfuscated and is not a real key

continuationToken :: Maybe Text

If ContinuationToken was sent with the request, it is included in the response.

httpStatus :: Int

The response's http status code.

newListObjectsVResponse

Arguments

:: Int
$sel:httpStatus:ListObjectsVResponse'

-> ListObjectsVResponse

Create a value of ListObjectsVResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyCount:ListObjectsVResponse', listObjectsVResponse_keyCount - KeyCount is the number of keys returned with this request. KeyCount will always be less than or equals to MaxKeys field. Say you ask for 50 keys, your result will include less than equals 50 keys

$sel:startAfter:ListObjectsV', listObjectsVResponse_startAfter - If StartAfter was sent with the request, it is included in the response.

$sel:commonPrefixes:ListObjectsVResponse', listObjectsVResponse_commonPrefixes - All of the keys (up to 1,000) rolled up into a common prefix count as a single return when calculating the number of returns.

A response can contain CommonPrefixes only if you specify a delimiter.

CommonPrefixes contains all (if there are any) keys between Prefix and the next occurrence of the string specified by a delimiter.

CommonPrefixes lists keys that act like subdirectories in the directory specified by Prefix.

For example, if the prefix is notes/ and the delimiter is a slash (/) as in notes/summer/july, the common prefix is notes/summer/. All of the keys that roll up into a common prefix count as a single return when calculating the number of returns.

$sel:delimiter:ListObjectsV', listObjectsVResponse_delimiter - Causes keys that contain the same string between the prefix and the first occurrence of the delimiter to be rolled up into a single result element in the CommonPrefixes collection. These rolled-up keys are not returned elsewhere in the response. Each rolled-up result counts as only one return against the MaxKeys value.

$sel:encodingType:ListObjectsV', listObjectsVResponse_encodingType - Encoding type used by Amazon S3 to encode object key names in the XML response.

If you specify the encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

Delimiter, Prefix, Key, and StartAfter.

$sel:prefix:ListObjectsV', listObjectsVResponse_prefix - Keys that begin with the indicated prefix.

$sel:maxKeys:ListObjectsV', listObjectsVResponse_maxKeys - Sets the maximum number of keys returned in the response. By default the action returns up to 1,000 key names. The response might contain fewer keys but will never contain more.

$sel:isTruncated:ListObjectsVResponse', listObjectsVResponse_isTruncated - Set to false if all of the results were returned. Set to true if more keys are available to return. If the number of results exceeds that specified by MaxKeys, all of the results might not be returned.

$sel:contents:ListObjectsVResponse', listObjectsVResponse_contents - Metadata about each object returned.

$sel:name:ListObjectsVResponse', listObjectsVResponse_name - The bucket name.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

$sel:nextContinuationToken:ListObjectsVResponse', listObjectsVResponse_nextContinuationToken - NextContinuationToken is sent when isTruncated is true, which means there are more keys in the bucket that can be listed. The next list requests to Amazon S3 can be continued with this NextContinuationToken. NextContinuationToken is obfuscated and is not a real key

$sel:continuationToken:ListObjectsV', listObjectsVResponse_continuationToken - If ContinuationToken was sent with the request, it is included in the response.

$sel:httpStatus:ListObjectsVResponse', listObjectsVResponse_httpStatus - The response's http status code.

Response Lenses

listObjectsVResponse_keyCount :: Lens' ListObjectsVResponse (Maybe Int)

KeyCount is the number of keys returned with this request. KeyCount will always be less than or equals to MaxKeys field. Say you ask for 50 keys, your result will include less than equals 50 keys

listObjectsVResponse_startAfter :: Lens' ListObjectsVResponse (Maybe Text)

If StartAfter was sent with the request, it is included in the response.

listObjectsVResponse_commonPrefixes :: Lens' ListObjectsVResponse (Maybe [CommonPrefix])

All of the keys (up to 1,000) rolled up into a common prefix count as a single return when calculating the number of returns.

A response can contain CommonPrefixes only if you specify a delimiter.

CommonPrefixes contains all (if there are any) keys between Prefix and the next occurrence of the string specified by a delimiter.

CommonPrefixes lists keys that act like subdirectories in the directory specified by Prefix.

For example, if the prefix is notes/ and the delimiter is a slash (/) as in notes/summer/july, the common prefix is notes/summer/. All of the keys that roll up into a common prefix count as a single return when calculating the number of returns.

listObjectsVResponse_delimiter :: Lens' ListObjectsVResponse (Maybe Delimiter)

Causes keys that contain the same string between the prefix and the first occurrence of the delimiter to be rolled up into a single result element in the CommonPrefixes collection. These rolled-up keys are not returned elsewhere in the response. Each rolled-up result counts as only one return against the MaxKeys value.

listObjectsVResponse_encodingType :: Lens' ListObjectsVResponse (Maybe EncodingType)

Encoding type used by Amazon S3 to encode object key names in the XML response.

If you specify the encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

Delimiter, Prefix, Key, and StartAfter.

listObjectsVResponse_prefix :: Lens' ListObjectsVResponse (Maybe Text)

Keys that begin with the indicated prefix.

listObjectsVResponse_maxKeys :: Lens' ListObjectsVResponse (Maybe Int)

Sets the maximum number of keys returned in the response. By default the action returns up to 1,000 key names. The response might contain fewer keys but will never contain more.

listObjectsVResponse_isTruncated :: Lens' ListObjectsVResponse (Maybe Bool)

Set to false if all of the results were returned. Set to true if more keys are available to return. If the number of results exceeds that specified by MaxKeys, all of the results might not be returned.

listObjectsVResponse_contents :: Lens' ListObjectsVResponse (Maybe [Object])

Metadata about each object returned.

listObjectsVResponse_name :: Lens' ListObjectsVResponse (Maybe BucketName)

The bucket name.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

listObjectsVResponse_nextContinuationToken :: Lens' ListObjectsVResponse (Maybe Text)

NextContinuationToken is sent when isTruncated is true, which means there are more keys in the bucket that can be listed. The next list requests to Amazon S3 can be continued with this NextContinuationToken. NextContinuationToken is obfuscated and is not a real key

listObjectsVResponse_continuationToken :: Lens' ListObjectsVResponse (Maybe Text)

If ContinuationToken was sent with the request, it is included in the response.

listObjectsVResponse_httpStatus :: Lens' ListObjectsVResponse Int

The response's http status code.