Network.AWS.S3.ListParts

Description

Lists the parts that have been uploaded for a specific multipart upload. This operation must include the upload ID, which you obtain by sending the initiate multipart upload request (see CreateMultipartUpload). This request returns a maximum of 1,000 uploaded parts. The default number of parts returned is 1,000 parts. You can restrict the number of parts returned by specifying the max-parts request parameter. If your multipart upload consists of more than 1,000 parts, the response returns an IsTruncated field with the value of true, and a NextPartNumberMarker element. In subsequent ListParts requests you can include the part-number-marker query string parameter and set its value to the NextPartNumberMarker field value from the previous response.

For more information on multipart uploads, see Uploading Objects Using Multipart Upload.

For information on permissions required to use the multipart upload API, see Multipart Upload and Permissions.

The following operations are related to ListParts:

CreateMultipartUpload

UploadPart

CompleteMultipartUpload

AbortMultipartUpload

ListMultipartUploads

This operation returns paginated results.

Creating a Request

data ListParts

See: newListParts smart constructor.

Constructors

ListParts'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

partNumberMarker :: Maybe Int

Specifies the part after which listing should begin. Only parts with higher part numbers will be listed.

maxParts :: Maybe Int

Sets the maximum number of parts to return.

requestPayer :: Maybe RequestPayer

bucket :: BucketName

The name of the bucket to which the parts are being uploaded.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

key :: ObjectKey

Object key for which the multipart upload was initiated.

uploadId :: Text

Upload ID identifying the multipart upload whose parts are being listed.

newListParts

Arguments

:: BucketName
$sel:bucket:ListParts'

-> ObjectKey
$sel:key:ListParts'

-> Text
$sel:uploadId:ListParts'

-> ListParts

Create a value of ListParts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:ListParts', listParts_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:partNumberMarker:ListParts', listParts_partNumberMarker - Specifies the part after which listing should begin. Only parts with higher part numbers will be listed.

$sel:maxParts:ListParts', listParts_maxParts - Sets the maximum number of parts to return.

$sel:requestPayer:ListParts', listParts_requestPayer - Undocumented member.

$sel:bucket:ListParts', listParts_bucket - The name of the bucket to which the parts are being uploaded.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

$sel:key:ListParts', listParts_key - Object key for which the multipart upload was initiated.

$sel:uploadId:ListParts', listParts_uploadId - Upload ID identifying the multipart upload whose parts are being listed.

Request Lenses

listParts_expectedBucketOwner :: Lens' ListParts (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

listParts_partNumberMarker :: Lens' ListParts (Maybe Int)

Specifies the part after which listing should begin. Only parts with higher part numbers will be listed.

listParts_maxParts :: Lens' ListParts (Maybe Int)

Sets the maximum number of parts to return.

listParts_requestPayer :: Lens' ListParts (Maybe RequestPayer)

Undocumented member.

listParts_bucket :: Lens' ListParts BucketName

The name of the bucket to which the parts are being uploaded.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

listParts_key :: Lens' ListParts ObjectKey

Object key for which the multipart upload was initiated.

listParts_uploadId :: Lens' ListParts Text

Upload ID identifying the multipart upload whose parts are being listed.

Destructuring the Response

data ListPartsResponse

See: newListPartsResponse smart constructor.

Constructors

ListPartsResponse'

Fields

key :: Maybe ObjectKey

Object key for which the multipart upload was initiated.

requestCharged :: Maybe RequestCharged

nextPartNumberMarker :: Maybe Int

When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

abortDate :: Maybe ISO8601

If the bucket has a lifecycle rule configured with an action to abort incomplete multipart uploads and the prefix in the lifecycle rule matches the object name in the request, then the response includes this header indicating when the initiated multipart upload will become eligible for abort operation. For more information, see Aborting Incomplete Multipart Uploads Using a Bucket Lifecycle Policy.

The response will also include the x-amz-abort-rule-id header that will provide the ID of the lifecycle configuration rule that defines this action.

uploadId :: Maybe Text

Upload ID identifying the multipart upload whose parts are being listed.

partNumberMarker :: Maybe Int

When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

maxParts :: Maybe Int

Maximum number of parts that were allowed in the response.

isTruncated :: Maybe Bool

Indicates whether the returned list of parts is truncated. A true value indicates that the list was truncated. A list can be truncated if the number of parts exceeds the limit returned in the MaxParts element.

storageClass :: Maybe StorageClass

Class of storage (STANDARD or REDUCED_REDUNDANCY) used to store the uploaded object.

parts :: Maybe [Part]

Container for elements related to a particular part. A response can contain zero or more Part elements.

abortRuleId :: Maybe Text

This header is returned along with the x-amz-abort-date header. It identifies applicable lifecycle configuration rule that defines the action to abort incomplete multipart uploads.

owner :: Maybe Owner

Container element that identifies the object owner, after the object is created. If multipart upload is initiated by an IAM user, this element provides the parent account ID and display name.

bucket :: Maybe BucketName

The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

initiator :: Maybe Initiator

Container element that identifies who initiated the multipart upload. If the initiator is an Amazon Web Services account, this element provides the same information as the Owner element. If the initiator is an IAM User, this element provides the user ARN and display name.

httpStatus :: Int

The response's http status code.

newListPartsResponse

Arguments

:: Int
$sel:httpStatus:ListPartsResponse'

-> ListPartsResponse

Create a value of ListPartsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:ListParts', listPartsResponse_key - Object key for which the multipart upload was initiated.

$sel:requestCharged:ListPartsResponse', listPartsResponse_requestCharged - Undocumented member.

$sel:nextPartNumberMarker:ListPartsResponse', listPartsResponse_nextPartNumberMarker - When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

$sel:abortDate:ListPartsResponse', listPartsResponse_abortDate - If the bucket has a lifecycle rule configured with an action to abort incomplete multipart uploads and the prefix in the lifecycle rule matches the object name in the request, then the response includes this header indicating when the initiated multipart upload will become eligible for abort operation. For more information, see Aborting Incomplete Multipart Uploads Using a Bucket Lifecycle Policy.

The response will also include the x-amz-abort-rule-id header that will provide the ID of the lifecycle configuration rule that defines this action.

$sel:uploadId:ListParts', listPartsResponse_uploadId - Upload ID identifying the multipart upload whose parts are being listed.

$sel:partNumberMarker:ListParts', listPartsResponse_partNumberMarker - When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

$sel:maxParts:ListParts', listPartsResponse_maxParts - Maximum number of parts that were allowed in the response.

$sel:isTruncated:ListPartsResponse', listPartsResponse_isTruncated - Indicates whether the returned list of parts is truncated. A true value indicates that the list was truncated. A list can be truncated if the number of parts exceeds the limit returned in the MaxParts element.

$sel:storageClass:ListPartsResponse', listPartsResponse_storageClass - Class of storage (STANDARD or REDUCED_REDUNDANCY) used to store the uploaded object.

$sel:parts:ListPartsResponse', listPartsResponse_parts - Container for elements related to a particular part. A response can contain zero or more Part elements.

$sel:abortRuleId:ListPartsResponse', listPartsResponse_abortRuleId - This header is returned along with the x-amz-abort-date header. It identifies applicable lifecycle configuration rule that defines the action to abort incomplete multipart uploads.

$sel:owner:ListPartsResponse', listPartsResponse_owner - Container element that identifies the object owner, after the object is created. If multipart upload is initiated by an IAM user, this element provides the parent account ID and display name.

$sel:bucket:ListParts', listPartsResponse_bucket - The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

$sel:initiator:ListPartsResponse', listPartsResponse_initiator - Container element that identifies who initiated the multipart upload. If the initiator is an Amazon Web Services account, this element provides the same information as the Owner element. If the initiator is an IAM User, this element provides the user ARN and display name.

$sel:httpStatus:ListPartsResponse', listPartsResponse_httpStatus - The response's http status code.

Response Lenses

listPartsResponse_key :: Lens' ListPartsResponse (Maybe ObjectKey)

Object key for which the multipart upload was initiated.

listPartsResponse_requestCharged :: Lens' ListPartsResponse (Maybe RequestCharged)

Undocumented member.

listPartsResponse_nextPartNumberMarker :: Lens' ListPartsResponse (Maybe Int)

When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

listPartsResponse_abortDate :: Lens' ListPartsResponse (Maybe UTCTime)

If the bucket has a lifecycle rule configured with an action to abort incomplete multipart uploads and the prefix in the lifecycle rule matches the object name in the request, then the response includes this header indicating when the initiated multipart upload will become eligible for abort operation. For more information, see Aborting Incomplete Multipart Uploads Using a Bucket Lifecycle Policy.

The response will also include the x-amz-abort-rule-id header that will provide the ID of the lifecycle configuration rule that defines this action.

listPartsResponse_uploadId :: Lens' ListPartsResponse (Maybe Text)

Upload ID identifying the multipart upload whose parts are being listed.

listPartsResponse_partNumberMarker :: Lens' ListPartsResponse (Maybe Int)

When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

listPartsResponse_maxParts :: Lens' ListPartsResponse (Maybe Int)

Maximum number of parts that were allowed in the response.

listPartsResponse_isTruncated :: Lens' ListPartsResponse (Maybe Bool)

Indicates whether the returned list of parts is truncated. A true value indicates that the list was truncated. A list can be truncated if the number of parts exceeds the limit returned in the MaxParts element.

listPartsResponse_storageClass :: Lens' ListPartsResponse (Maybe StorageClass)

Class of storage (STANDARD or REDUCED_REDUNDANCY) used to store the uploaded object.

listPartsResponse_parts :: Lens' ListPartsResponse (Maybe [Part])

Container for elements related to a particular part. A response can contain zero or more Part elements.

listPartsResponse_abortRuleId :: Lens' ListPartsResponse (Maybe Text)

This header is returned along with the x-amz-abort-date header. It identifies applicable lifecycle configuration rule that defines the action to abort incomplete multipart uploads.

listPartsResponse_owner :: Lens' ListPartsResponse (Maybe Owner)

Container element that identifies the object owner, after the object is created. If multipart upload is initiated by an IAM user, this element provides the parent account ID and display name.

listPartsResponse_bucket :: Lens' ListPartsResponse (Maybe BucketName)

The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

listPartsResponse_initiator :: Lens' ListPartsResponse (Maybe Initiator)

Container element that identifies who initiated the multipart upload. If the initiator is an Amazon Web Services account, this element provides the same information as the Owner element. If the initiator is an IAM User, this element provides the user ARN and display name.

listPartsResponse_httpStatus :: Lens' ListPartsResponse Int

The response's http status code.