Network.AWS.S3.PutBucketAccelerateConfiguration

Description

Sets the accelerate configuration of an existing bucket. Amazon S3 Transfer Acceleration is a bucket-level feature that enables you to perform faster data transfers to Amazon S3.

To use this operation, you must have permission to perform the s3:PutAccelerateConfiguration action. The bucket owner has this permission by default. The bucket owner can grant this permission to others. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to Your Amazon S3 Resources.

The Transfer Acceleration state of a bucket can be set to one of the following two values:

Enabled – Enables accelerated data transfers to the bucket.

Suspended – Disables accelerated data transfers to the bucket.

The GetBucketAccelerateConfiguration action returns the transfer acceleration state of a bucket.

After setting the Transfer Acceleration state of a bucket to Enabled, it might take up to thirty minutes before the data transfer rates to the bucket increase.

The name of the bucket used for Transfer Acceleration must be DNS-compliant and must not contain periods (".").

For more information about transfer acceleration, see Transfer Acceleration.

The following operations are related to PutBucketAccelerateConfiguration:

GetBucketAccelerateConfiguration

CreateBucket

Creating a Request

data PutBucketAccelerateConfiguration

See: newPutBucketAccelerateConfiguration smart constructor.

Constructors

PutBucketAccelerateConfiguration'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The name of the bucket for which the accelerate configuration is set.

accelerateConfiguration :: AccelerateConfiguration

Container for setting the transfer acceleration state.

newPutBucketAccelerateConfiguration

Arguments

:: BucketName
$sel:bucket:PutBucketAccelerateConfiguration'

-> AccelerateConfiguration
$sel:accelerateConfiguration:PutBucketAccelerateConfiguration'

-> PutBucketAccelerateConfiguration

Create a value of PutBucketAccelerateConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:PutBucketAccelerateConfiguration', putBucketAccelerateConfiguration_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:PutBucketAccelerateConfiguration', putBucketAccelerateConfiguration_bucket - The name of the bucket for which the accelerate configuration is set.

$sel:accelerateConfiguration:PutBucketAccelerateConfiguration', putBucketAccelerateConfiguration_accelerateConfiguration - Container for setting the transfer acceleration state.

Request Lenses

putBucketAccelerateConfiguration_expectedBucketOwner :: Lens' PutBucketAccelerateConfiguration (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putBucketAccelerateConfiguration_bucket :: Lens' PutBucketAccelerateConfiguration BucketName

The name of the bucket for which the accelerate configuration is set.

putBucketAccelerateConfiguration_accelerateConfiguration :: Lens' PutBucketAccelerateConfiguration AccelerateConfiguration

Container for setting the transfer acceleration state.

Destructuring the Response

data PutBucketAccelerateConfigurationResponse

See: newPutBucketAccelerateConfigurationResponse smart constructor.

Constructors

PutBucketAccelerateConfigurationResponse'

newPutBucketAccelerateConfigurationResponse :: PutBucketAccelerateConfigurationResponse

Create a value of PutBucketAccelerateConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.