Network.AWS.S3.PutBucketAcl

Description

Sets the permissions on an existing bucket using access control lists (ACL). For more information, see Using ACLs. To set the ACL of a bucket, you must have WRITE_ACP permission.

You can use one of the following two ways to set a bucket's permissions:

Specify the ACL in the request body

Specify permissions using request headers

You cannot specify access permission using both the body and the request headers.

Depending on your application needs, you may choose to set the ACL on a bucket using either the request body or the headers. For example, if you have an existing application that updates a bucket ACL using the request body, then you can continue to use that approach.

Access Permissions

You can set access permissions using one of the following methods:

Specify a canned ACL with the x-amz-acl request header. Amazon S3 supports a set of predefined ACLs, known as canned ACLs. Each canned ACL has a predefined set of grantees and permissions. Specify the canned ACL name as the value of x-amz-acl. If you use this header, you cannot use other access control-specific headers in your request. For more information, see Canned ACL.

Specify access permissions explicitly with the x-amz-grant-read, x-amz-grant-read-acp, x-amz-grant-write-acp, and x-amz-grant-full-control headers. When using these headers, you specify explicit access permissions and grantees (Amazon Web Services accounts or Amazon S3 groups) who will receive the permission. If you use these ACL-specific headers, you cannot use the x-amz-acl header to set a canned ACL. These parameters map to the set of permissions that Amazon S3 supports in an ACL. For more information, see Access Control List (ACL) Overview.

You specify each grantee as a type=value pair, where the type is one of the following:

id – if the value specified is the canonical user ID of an Amazon Web Services account

uri – if you are granting permissions to a predefined group

emailAddress – if the value specified is the email address of an Amazon Web Services account

Using email addresses to specify a grantee is only supported in the following Amazon Web Services Regions:

US East (N. Virginia)

US West (N. California)

US West (Oregon)

Asia Pacific (Singapore)

Asia Pacific (Sydney)

Asia Pacific (Tokyo)

Europe (Ireland)

South America (São Paulo)

For a list of all the Amazon S3 supported Regions and endpoints, see Regions and Endpoints in the Amazon Web Services General Reference.

For example, the following x-amz-grant-write header grants create, overwrite, and delete objects permission to LogDelivery group predefined by Amazon S3 and two Amazon Web Services accounts identified by their email addresses.

x-amz-grant-write: uri="http://acs.amazonaws.com/groups/s3/LogDelivery", id="111122223333", id="555566667777"

You can use either a canned ACL or specify access permissions explicitly. You cannot do both.

Grantee Values

You can specify the person (grantee) to whom you're assigning access rights (using request elements) in the following ways:

By the person's ID:

<Grantee xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:type="CanonicalUser"><ID><>ID<></ID><DisplayName><>GranteesEmail<></DisplayName> </Grantee>

DisplayName is optional and ignored in the request

By URI:

<Grantee xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:type="Group"><URI><>http://acs.amazonaws.com/groups/global/AuthenticatedUsers<></URI></Grantee>

By Email address:

<Grantee xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:type="AmazonCustomerByEmail"><EmailAddress><>Grantees@email.com<></EmailAddress>lt;/Grantee>

The grantee is resolved to the CanonicalUser and, in a response to a GET Object acl request, appears as the CanonicalUser.

Using email addresses to specify a grantee is only supported in the following Amazon Web Services Regions:

US East (N. Virginia)

US West (N. California)

US West (Oregon)

Asia Pacific (Singapore)

Asia Pacific (Sydney)

Asia Pacific (Tokyo)

Europe (Ireland)

South America (São Paulo)

For a list of all the Amazon S3 supported Regions and endpoints, see Regions and Endpoints in the Amazon Web Services General Reference.

Related Resources

CreateBucket

DeleteBucket

GetObjectAcl

Creating a Request

data PutBucketAcl

See: newPutBucketAcl smart constructor.

Constructors

PutBucketAcl'

Fields

grantRead :: Maybe Text

Allows grantee to list the objects in the bucket.

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

accessControlPolicy :: Maybe AccessControlPolicy

Contains the elements that set the ACL permissions for an object per grantee.

contentMD5 :: Maybe Text

The base64-encoded 128-bit MD5 digest of the data. This header must be used as a message integrity check to verify that the request body was not corrupted in transit. For more information, go to RFC 1864.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

grantWriteACP :: Maybe Text

Allows grantee to write the ACL for the applicable bucket.

acl :: Maybe BucketCannedACL

The canned ACL to apply to the bucket.

grantReadACP :: Maybe Text

Allows grantee to read the bucket ACL.

grantWrite :: Maybe Text

Allows grantee to create new objects in the bucket.

For the bucket and object owners of existing objects, also allows deletions and overwrites of those objects.

grantFullControl :: Maybe Text

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

bucket :: BucketName

The bucket to which to apply the ACL.

newPutBucketAcl

Arguments

:: BucketName
$sel:bucket:PutBucketAcl'

-> PutBucketAcl

Create a value of PutBucketAcl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:grantRead:PutBucketAcl', putBucketAcl_grantRead - Allows grantee to list the objects in the bucket.

$sel:expectedBucketOwner:PutBucketAcl', putBucketAcl_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:accessControlPolicy:PutBucketAcl', putBucketAcl_accessControlPolicy - Contains the elements that set the ACL permissions for an object per grantee.

$sel:contentMD5:PutBucketAcl', putBucketAcl_contentMD5 - The base64-encoded 128-bit MD5 digest of the data. This header must be used as a message integrity check to verify that the request body was not corrupted in transit. For more information, go to RFC 1864.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

$sel:grantWriteACP:PutBucketAcl', putBucketAcl_grantWriteACP - Allows grantee to write the ACL for the applicable bucket.

$sel:acl:PutBucketAcl', putBucketAcl_acl - The canned ACL to apply to the bucket.

$sel:grantReadACP:PutBucketAcl', putBucketAcl_grantReadACP - Allows grantee to read the bucket ACL.

$sel:grantWrite:PutBucketAcl', putBucketAcl_grantWrite - Allows grantee to create new objects in the bucket.

For the bucket and object owners of existing objects, also allows deletions and overwrites of those objects.

$sel:grantFullControl:PutBucketAcl', putBucketAcl_grantFullControl - Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

$sel:bucket:PutBucketAcl', putBucketAcl_bucket - The bucket to which to apply the ACL.

Request Lenses

putBucketAcl_grantRead :: Lens' PutBucketAcl (Maybe Text)

Allows grantee to list the objects in the bucket.

putBucketAcl_expectedBucketOwner :: Lens' PutBucketAcl (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putBucketAcl_accessControlPolicy :: Lens' PutBucketAcl (Maybe AccessControlPolicy)

Contains the elements that set the ACL permissions for an object per grantee.

putBucketAcl_contentMD5 :: Lens' PutBucketAcl (Maybe Text)

The base64-encoded 128-bit MD5 digest of the data. This header must be used as a message integrity check to verify that the request body was not corrupted in transit. For more information, go to RFC 1864.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

putBucketAcl_grantWriteACP :: Lens' PutBucketAcl (Maybe Text)

Allows grantee to write the ACL for the applicable bucket.

putBucketAcl_acl :: Lens' PutBucketAcl (Maybe BucketCannedACL)

The canned ACL to apply to the bucket.

putBucketAcl_grantReadACP :: Lens' PutBucketAcl (Maybe Text)

Allows grantee to read the bucket ACL.

putBucketAcl_grantWrite :: Lens' PutBucketAcl (Maybe Text)

Allows grantee to create new objects in the bucket.

For the bucket and object owners of existing objects, also allows deletions and overwrites of those objects.

putBucketAcl_grantFullControl :: Lens' PutBucketAcl (Maybe Text)

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

putBucketAcl_bucket :: Lens' PutBucketAcl BucketName

The bucket to which to apply the ACL.

Destructuring the Response

data PutBucketAclResponse

See: newPutBucketAclResponse smart constructor.

Constructors

PutBucketAclResponse'

newPutBucketAclResponse :: PutBucketAclResponse

Create a value of PutBucketAclResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.