Network.AWS.S3.PutBucketCors

Description

Sets the cors configuration for your bucket. If the configuration exists, Amazon S3 replaces it.

To use this operation, you must be allowed to perform the s3:PutBucketCORS action. By default, the bucket owner has this permission and can grant it to others.

You set this configuration on a bucket so that the bucket can service cross-origin requests. For example, you might want to enable a request whose origin is http://www.example.com to access your Amazon S3 bucket at my.example.bucket.com by using the browser's XMLHttpRequest capability.

To enable cross-origin resource sharing (CORS) on a bucket, you add the cors subresource to the bucket. The cors subresource is an XML document in which you configure rules that identify origins and the HTTP methods that can be executed on your bucket. The document is limited to 64 KB in size.

When Amazon S3 receives a cross-origin request (or a pre-flight OPTIONS request) against a bucket, it evaluates the cors configuration on the bucket and uses the first CORSRule rule that matches the incoming browser request to enable a cross-origin request. For a rule to match, the following conditions must be met:

The request's Origin header must match AllowedOrigin elements.

The request method (for example, GET, PUT, HEAD, and so on) or the Access-Control-Request-Method header in case of a pre-flight OPTIONS request must be one of the AllowedMethod elements.

Every header specified in the Access-Control-Request-Headers request header of a pre-flight request must match an AllowedHeader element.

For more information about CORS, go to Enabling Cross-Origin Resource Sharing in the Amazon S3 User Guide.

Related Resources

GetBucketCors

DeleteBucketCors

RESTOPTIONSobject

Creating a Request

data PutBucketCors

See: newPutBucketCors smart constructor.

Constructors

PutBucketCors'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

contentMD5 :: Maybe Text

The base64-encoded 128-bit MD5 digest of the data. This header must be used as a message integrity check to verify that the request body was not corrupted in transit. For more information, go to RFC 1864.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

bucket :: BucketName

Specifies the bucket impacted by the corsconfiguration.

cORSConfiguration :: CORSConfiguration

Describes the cross-origin access configuration for objects in an Amazon S3 bucket. For more information, see Enabling Cross-Origin Resource Sharing in the Amazon S3 User Guide.

newPutBucketCors

Arguments

:: BucketName
$sel:bucket:PutBucketCors'

-> CORSConfiguration
$sel:cORSConfiguration:PutBucketCors'

-> PutBucketCors

Create a value of PutBucketCors with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:PutBucketCors', putBucketCors_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:contentMD5:PutBucketCors', putBucketCors_contentMD5 - The base64-encoded 128-bit MD5 digest of the data. This header must be used as a message integrity check to verify that the request body was not corrupted in transit. For more information, go to RFC 1864.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

$sel:bucket:PutBucketCors', putBucketCors_bucket - Specifies the bucket impacted by the corsconfiguration.

$sel:cORSConfiguration:PutBucketCors', putBucketCors_cORSConfiguration - Describes the cross-origin access configuration for objects in an Amazon S3 bucket. For more information, see Enabling Cross-Origin Resource Sharing in the Amazon S3 User Guide.

Request Lenses

putBucketCors_expectedBucketOwner :: Lens' PutBucketCors (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putBucketCors_contentMD5 :: Lens' PutBucketCors (Maybe Text)

The base64-encoded 128-bit MD5 digest of the data. This header must be used as a message integrity check to verify that the request body was not corrupted in transit. For more information, go to RFC 1864.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

putBucketCors_bucket :: Lens' PutBucketCors BucketName

Specifies the bucket impacted by the corsconfiguration.

putBucketCors_cORSConfiguration :: Lens' PutBucketCors CORSConfiguration

Describes the cross-origin access configuration for objects in an Amazon S3 bucket. For more information, see Enabling Cross-Origin Resource Sharing in the Amazon S3 User Guide.

Destructuring the Response

data PutBucketCorsResponse

See: newPutBucketCorsResponse smart constructor.

Constructors

PutBucketCorsResponse'

newPutBucketCorsResponse :: PutBucketCorsResponse

Create a value of PutBucketCorsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.