Network.AWS.S3.PutBucketEncryption

Description

This action uses the encryption subresource to configure default encryption and Amazon S3 Bucket Key for an existing bucket.

Default encryption for a bucket can use server-side encryption with Amazon S3-managed keys (SSE-S3) or customer managed keys (SSE-KMS). If you specify default encryption using SSE-KMS, you can also configure Amazon S3 Bucket Key. For information about default encryption, see Amazon S3 default bucket encryption in the Amazon S3 User Guide. For more information about S3 Bucket Keys, see Amazon S3 Bucket Keys in the Amazon S3 User Guide.

This action requires Amazon Web Services Signature Version 4. For more information, see Authenticating Requests (Amazon Web Services Signature Version 4).

To use this operation, you must have permissions to perform the s3:PutEncryptionConfiguration action. The bucket owner has this permission by default. The bucket owner can grant this permission to others. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to Your Amazon S3 Resources in the Amazon S3 User Guide.

Related Resources

GetBucketEncryption

DeleteBucketEncryption

Creating a Request

data PutBucketEncryption

See: newPutBucketEncryption smart constructor.

Constructors

PutBucketEncryption'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

contentMD5 :: Maybe Text

The base64-encoded 128-bit MD5 digest of the server-side encryption configuration.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

bucket :: BucketName

Specifies default encryption for a bucket using server-side encryption with Amazon S3-managed keys (SSE-S3) or customer managed keys (SSE-KMS). For information about the Amazon S3 default encryption feature, see Amazon S3 Default Bucket Encryption in the Amazon S3 User Guide.

serverSideEncryptionConfiguration :: ServerSideEncryptionConfiguration

newPutBucketEncryption

Arguments

:: BucketName
$sel:bucket:PutBucketEncryption'

-> ServerSideEncryptionConfiguration
$sel:serverSideEncryptionConfiguration:PutBucketEncryption'

-> PutBucketEncryption

Create a value of PutBucketEncryption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:PutBucketEncryption', putBucketEncryption_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:contentMD5:PutBucketEncryption', putBucketEncryption_contentMD5 - The base64-encoded 128-bit MD5 digest of the server-side encryption configuration.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

$sel:bucket:PutBucketEncryption', putBucketEncryption_bucket - Specifies default encryption for a bucket using server-side encryption with Amazon S3-managed keys (SSE-S3) or customer managed keys (SSE-KMS). For information about the Amazon S3 default encryption feature, see Amazon S3 Default Bucket Encryption in the Amazon S3 User Guide.

$sel:serverSideEncryptionConfiguration:PutBucketEncryption', putBucketEncryption_serverSideEncryptionConfiguration - Undocumented member.

Request Lenses

putBucketEncryption_expectedBucketOwner :: Lens' PutBucketEncryption (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putBucketEncryption_contentMD5 :: Lens' PutBucketEncryption (Maybe Text)

The base64-encoded 128-bit MD5 digest of the server-side encryption configuration.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

putBucketEncryption_bucket :: Lens' PutBucketEncryption BucketName

Specifies default encryption for a bucket using server-side encryption with Amazon S3-managed keys (SSE-S3) or customer managed keys (SSE-KMS). For information about the Amazon S3 default encryption feature, see Amazon S3 Default Bucket Encryption in the Amazon S3 User Guide.

putBucketEncryption_serverSideEncryptionConfiguration :: Lens' PutBucketEncryption ServerSideEncryptionConfiguration

Undocumented member.

Destructuring the Response

data PutBucketEncryptionResponse

See: newPutBucketEncryptionResponse smart constructor.

Constructors

PutBucketEncryptionResponse'

newPutBucketEncryptionResponse :: PutBucketEncryptionResponse

Create a value of PutBucketEncryptionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.