Network.AWS.S3.PutBucketInventoryConfiguration

Description

This implementation of the PUT action adds an inventory configuration (identified by the inventory ID) to the bucket. You can have up to 1,000 inventory configurations per bucket.

Amazon S3 inventory generates inventories of the objects in the bucket on a daily or weekly basis, and the results are published to a flat file. The bucket that is inventoried is called the source bucket, and the bucket where the inventory flat file is stored is called the destination bucket. The destination bucket must be in the same Amazon Web Services Region as the source bucket.

When you configure an inventory for a source bucket, you specify the destination bucket where you want the inventory to be stored, and whether to generate the inventory daily or weekly. You can also configure what object metadata to include and whether to inventory all object versions or only current versions. For more information, see Amazon S3 Inventory in the Amazon S3 User Guide.

You must create a bucket policy on the destination bucket to grant permissions to Amazon S3 to write objects to the bucket in the defined location. For an example policy, see Granting Permissions for Amazon S3 Inventory and Storage Class Analysis.

To use this operation, you must have permissions to perform the s3:PutInventoryConfiguration action. The bucket owner has this permission by default and can grant this permission to others. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to Your Amazon S3 Resources in the Amazon S3 User Guide.

Special Errors

HTTP 400 Bad Request Error

Code: InvalidArgument

Cause: Invalid Argument

HTTP 400 Bad Request Error

Code: TooManyConfigurations

Cause: You are attempting to create a new configuration but have already reached the 1,000-configuration limit.

HTTP 403 Forbidden Error

Code: AccessDenied

Cause: You are not the owner of the specified bucket, or you do not have the s3:PutInventoryConfiguration bucket permission to set the configuration on the bucket.

Related Resources

GetBucketInventoryConfiguration

DeleteBucketInventoryConfiguration

ListBucketInventoryConfigurations

Creating a Request

data PutBucketInventoryConfiguration

See: newPutBucketInventoryConfiguration smart constructor.

Constructors

PutBucketInventoryConfiguration'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The name of the bucket where the inventory configuration will be stored.

id :: Text

The ID used to identify the inventory configuration.

inventoryConfiguration :: InventoryConfiguration

Specifies the inventory configuration.

newPutBucketInventoryConfiguration

Arguments

:: BucketName
$sel:bucket:PutBucketInventoryConfiguration'

-> Text
$sel:id:PutBucketInventoryConfiguration'

-> InventoryConfiguration
$sel:inventoryConfiguration:PutBucketInventoryConfiguration'

-> PutBucketInventoryConfiguration

Create a value of PutBucketInventoryConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:PutBucketInventoryConfiguration', putBucketInventoryConfiguration_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:PutBucketInventoryConfiguration', putBucketInventoryConfiguration_bucket - The name of the bucket where the inventory configuration will be stored.

$sel:id:PutBucketInventoryConfiguration', putBucketInventoryConfiguration_id - The ID used to identify the inventory configuration.

$sel:inventoryConfiguration:PutBucketInventoryConfiguration', putBucketInventoryConfiguration_inventoryConfiguration - Specifies the inventory configuration.

Request Lenses

putBucketInventoryConfiguration_expectedBucketOwner :: Lens' PutBucketInventoryConfiguration (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putBucketInventoryConfiguration_bucket :: Lens' PutBucketInventoryConfiguration BucketName

The name of the bucket where the inventory configuration will be stored.

putBucketInventoryConfiguration_id :: Lens' PutBucketInventoryConfiguration Text

The ID used to identify the inventory configuration.

putBucketInventoryConfiguration_inventoryConfiguration :: Lens' PutBucketInventoryConfiguration InventoryConfiguration

Specifies the inventory configuration.

Destructuring the Response

data PutBucketInventoryConfigurationResponse

See: newPutBucketInventoryConfigurationResponse smart constructor.

Constructors

PutBucketInventoryConfigurationResponse'

newPutBucketInventoryConfigurationResponse :: PutBucketInventoryConfigurationResponse

Create a value of PutBucketInventoryConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.