Network.AWS.S3.PutBucketNotificationConfiguration

Description

Enables notifications of specified events for a bucket. For more information about event notifications, see Configuring Event Notifications.

Using this API, you can replace an existing notification configuration. The configuration is an XML file that defines the event types that you want Amazon S3 to publish and the destination where you want Amazon S3 to publish an event notification when it detects an event of the specified type.

By default, your bucket has no event notifications configured. That is, the notification configuration will be an empty NotificationConfiguration.

<NotificationConfiguration>

</NotificationConfiguration>

This action replaces the existing notification configuration with the configuration you include in the request body.

After Amazon S3 receives this request, it first verifies that any Amazon Simple Notification Service (Amazon SNS) or Amazon Simple Queue Service (Amazon SQS) destination exists, and that the bucket owner has permission to publish to it by sending a test notification. In the case of Lambda destinations, Amazon S3 verifies that the Lambda function permissions grant Amazon S3 permission to invoke the function from the Amazon S3 bucket. For more information, see Configuring Notifications for Amazon S3 Events.

You can disable notifications by adding the empty NotificationConfiguration element.

By default, only the bucket owner can configure notifications on a bucket. However, bucket owners can use a bucket policy to grant permission to other users to set this configuration with s3:PutBucketNotification permission.

The PUT notification is an atomic operation. For example, suppose your notification configuration includes SNS topic, SQS queue, and Lambda function configurations. When you send a PUT request with this configuration, Amazon S3 sends test messages to your SNS topic. If the message fails, the entire PUT action will fail, and Amazon S3 will not add the configuration to your bucket.

Responses

If the configuration in the request body includes only one TopicConfiguration specifying only the s3:ReducedRedundancyLostObject event type, the response will also include the x-amz-sns-test-message-id header containing the message ID of the test notification sent to the topic.

The following action is related to PutBucketNotificationConfiguration:

GetBucketNotificationConfiguration

Creating a Request

data PutBucketNotificationConfiguration

See: newPutBucketNotificationConfiguration smart constructor.

Constructors

PutBucketNotificationConfiguration'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The name of the bucket.

notificationConfiguration :: NotificationConfiguration

newPutBucketNotificationConfiguration

Arguments

:: BucketName
$sel:bucket:PutBucketNotificationConfiguration'

-> NotificationConfiguration
$sel:notificationConfiguration:PutBucketNotificationConfiguration'

-> PutBucketNotificationConfiguration

Create a value of PutBucketNotificationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:PutBucketNotificationConfiguration', putBucketNotificationConfiguration_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:PutBucketNotificationConfiguration', putBucketNotificationConfiguration_bucket - The name of the bucket.

$sel:notificationConfiguration:PutBucketNotificationConfiguration', putBucketNotificationConfiguration_notificationConfiguration - Undocumented member.

Request Lenses

putBucketNotificationConfiguration_expectedBucketOwner :: Lens' PutBucketNotificationConfiguration (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putBucketNotificationConfiguration_bucket :: Lens' PutBucketNotificationConfiguration BucketName

The name of the bucket.

putBucketNotificationConfiguration_notificationConfiguration :: Lens' PutBucketNotificationConfiguration NotificationConfiguration

Undocumented member.

Destructuring the Response

data PutBucketNotificationConfigurationResponse

See: newPutBucketNotificationConfigurationResponse smart constructor.

Constructors

PutBucketNotificationConfigurationResponse'

newPutBucketNotificationConfigurationResponse :: PutBucketNotificationConfigurationResponse

Create a value of PutBucketNotificationConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.