Network.AWS.S3.PutBucketPolicy

Description

Applies an Amazon S3 bucket policy to an Amazon S3 bucket. If you are using an identity other than the root user of the Amazon Web Services account that owns the bucket, the calling identity must have the PutBucketPolicy permissions on the specified bucket and belong to the bucket owner's account in order to use this operation.

If you don't have PutBucketPolicy permissions, Amazon S3 returns a 403 Access Denied error. If you have the correct permissions, but you're not using an identity that belongs to the bucket owner's account, Amazon S3 returns a 405 Method Not Allowed error.

As a security precaution, the root user of the Amazon Web Services account that owns a bucket can always use this operation, even if the policy explicitly denies the root user the ability to perform this action.

For more information, see Bucket policy examples.

The following operations are related to PutBucketPolicy:

CreateBucket

DeleteBucket

Creating a Request

data PutBucketPolicy

See: newPutBucketPolicy smart constructor.

Constructors

PutBucketPolicy'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

contentMD5 :: Maybe Text

The MD5 hash of the request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

confirmRemoveSelfBucketAccess :: Maybe Bool

Set this parameter to true to confirm that you want to remove your permissions to change this bucket policy in the future.

bucket :: BucketName

The name of the bucket.

policy :: ByteString

The bucket policy as a JSON document.

newPutBucketPolicy

Arguments

:: BucketName
$sel:bucket:PutBucketPolicy'

-> ByteString
$sel:policy:PutBucketPolicy'

-> PutBucketPolicy

Create a value of PutBucketPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:PutBucketPolicy', putBucketPolicy_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:contentMD5:PutBucketPolicy', putBucketPolicy_contentMD5 - The MD5 hash of the request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

$sel:confirmRemoveSelfBucketAccess:PutBucketPolicy', putBucketPolicy_confirmRemoveSelfBucketAccess - Set this parameter to true to confirm that you want to remove your permissions to change this bucket policy in the future.

$sel:bucket:PutBucketPolicy', putBucketPolicy_bucket - The name of the bucket.

$sel:policy:PutBucketPolicy', putBucketPolicy_policy - The bucket policy as a JSON document.

Request Lenses

putBucketPolicy_expectedBucketOwner :: Lens' PutBucketPolicy (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putBucketPolicy_contentMD5 :: Lens' PutBucketPolicy (Maybe Text)

The MD5 hash of the request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

putBucketPolicy_confirmRemoveSelfBucketAccess :: Lens' PutBucketPolicy (Maybe Bool)

Set this parameter to true to confirm that you want to remove your permissions to change this bucket policy in the future.

putBucketPolicy_bucket :: Lens' PutBucketPolicy BucketName

The name of the bucket.

putBucketPolicy_policy :: Lens' PutBucketPolicy ByteString

The bucket policy as a JSON document.

Destructuring the Response

data PutBucketPolicyResponse

See: newPutBucketPolicyResponse smart constructor.

Constructors

PutBucketPolicyResponse'

newPutBucketPolicyResponse :: PutBucketPolicyResponse

Create a value of PutBucketPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.