Network.AWS.S3.PutBucketReplication

Description

Creates a replication configuration or replaces an existing one. For more information, see Replication in the Amazon S3 User Guide.

Specify the replication configuration in the request body. In the replication configuration, you provide the name of the destination bucket or buckets where you want Amazon S3 to replicate objects, the IAM role that Amazon S3 can assume to replicate objects on your behalf, and other relevant information.

A replication configuration must include at least one rule, and can contain a maximum of 1,000. Each rule identifies a subset of objects to replicate by filtering the objects in the source bucket. To choose additional subsets of objects to replicate, add a rule for each subset.

To specify a subset of the objects in the source bucket to apply a replication rule to, add the Filter element as a child of the Rule element. You can filter objects based on an object key prefix, one or more object tags, or both. When you add the Filter element in the configuration, you must also add the following elements: DeleteMarkerReplication, Status, and Priority.

If you are using an earlier version of the replication configuration, Amazon S3 handles replication of delete markers differently. For more information, see Backward Compatibility.

For information about enabling versioning on a bucket, see Using Versioning.

Handling Replication of Encrypted Objects

By default, Amazon S3 doesn't replicate objects that are stored at rest using server-side encryption with KMS keys. To replicate Amazon Web Services KMS-encrypted objects, add the following: SourceSelectionCriteria, SseKmsEncryptedObjects, Status, EncryptionConfiguration, and ReplicaKmsKeyID. For information about replication configuration, see Replicating Objects Created with SSE Using KMS keys.

For information on PutBucketReplication errors, see List of replication-related error codes

Permissions

To create a PutBucketReplication request, you must have s3:PutReplicationConfiguration permissions for the bucket.

By default, a resource owner, in this case the Amazon Web Services account that created the bucket, can perform this operation. The resource owner can also grant others permissions to perform the operation. For more information about permissions, see Specifying Permissions in a Policy and Managing Access Permissions to Your Amazon S3 Resources.

To perform this operation, the user or role performing the action must have the iam:PassRole permission.

The following operations are related to PutBucketReplication:

GetBucketReplication

DeleteBucketReplication

Creating a Request

data PutBucketReplication

See: newPutBucketReplication smart constructor.

Constructors

PutBucketReplication'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

contentMD5 :: Maybe Text

The base64-encoded 128-bit MD5 digest of the data. You must use this header as a message integrity check to verify that the request body was not corrupted in transit. For more information, see RFC 1864.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

token :: Maybe Text

A token to allow Object Lock to be enabled for an existing bucket.

bucket :: BucketName

The name of the bucket

replicationConfiguration :: ReplicationConfiguration

newPutBucketReplication

Arguments

:: BucketName
$sel:bucket:PutBucketReplication'

-> ReplicationConfiguration
$sel:replicationConfiguration:PutBucketReplication'

-> PutBucketReplication

Create a value of PutBucketReplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:PutBucketReplication', putBucketReplication_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:contentMD5:PutBucketReplication', putBucketReplication_contentMD5 - The base64-encoded 128-bit MD5 digest of the data. You must use this header as a message integrity check to verify that the request body was not corrupted in transit. For more information, see RFC 1864.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

$sel:token:PutBucketReplication', putBucketReplication_token - A token to allow Object Lock to be enabled for an existing bucket.

$sel:bucket:PutBucketReplication', putBucketReplication_bucket - The name of the bucket

$sel:replicationConfiguration:PutBucketReplication', putBucketReplication_replicationConfiguration - Undocumented member.

Request Lenses

putBucketReplication_expectedBucketOwner :: Lens' PutBucketReplication (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putBucketReplication_contentMD5 :: Lens' PutBucketReplication (Maybe Text)

The base64-encoded 128-bit MD5 digest of the data. You must use this header as a message integrity check to verify that the request body was not corrupted in transit. For more information, see RFC 1864.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

putBucketReplication_token :: Lens' PutBucketReplication (Maybe Text)

A token to allow Object Lock to be enabled for an existing bucket.

putBucketReplication_bucket :: Lens' PutBucketReplication BucketName

The name of the bucket

putBucketReplication_replicationConfiguration :: Lens' PutBucketReplication ReplicationConfiguration

Undocumented member.

Destructuring the Response

data PutBucketReplicationResponse

See: newPutBucketReplicationResponse smart constructor.

Constructors

PutBucketReplicationResponse'

newPutBucketReplicationResponse :: PutBucketReplicationResponse

Create a value of PutBucketReplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.