Network.AWS.S3.PutBucketRequestPayment

Description

Sets the request payment configuration for a bucket. By default, the bucket owner pays for downloads from the bucket. This configuration parameter enables the bucket owner (only) to specify that the person requesting the download will be charged for the download. For more information, see Requester Pays Buckets.

The following operations are related to PutBucketRequestPayment:

CreateBucket

GetBucketRequestPayment

Creating a Request

data PutBucketRequestPayment

See: newPutBucketRequestPayment smart constructor.

Constructors

PutBucketRequestPayment'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

contentMD5 :: Maybe Text

The base64-encoded 128-bit MD5 digest of the data. You must use this header as a message integrity check to verify that the request body was not corrupted in transit. For more information, see RFC 1864.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

bucket :: BucketName

The bucket name.

requestPaymentConfiguration :: RequestPaymentConfiguration

Container for Payer.

newPutBucketRequestPayment

Arguments

:: BucketName
$sel:bucket:PutBucketRequestPayment'

-> RequestPaymentConfiguration
$sel:requestPaymentConfiguration:PutBucketRequestPayment'

-> PutBucketRequestPayment

Create a value of PutBucketRequestPayment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:PutBucketRequestPayment', putBucketRequestPayment_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:contentMD5:PutBucketRequestPayment', putBucketRequestPayment_contentMD5 - The base64-encoded 128-bit MD5 digest of the data. You must use this header as a message integrity check to verify that the request body was not corrupted in transit. For more information, see RFC 1864.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

$sel:bucket:PutBucketRequestPayment', putBucketRequestPayment_bucket - The bucket name.

$sel:requestPaymentConfiguration:PutBucketRequestPayment', putBucketRequestPayment_requestPaymentConfiguration - Container for Payer.

Request Lenses

putBucketRequestPayment_expectedBucketOwner :: Lens' PutBucketRequestPayment (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putBucketRequestPayment_contentMD5 :: Lens' PutBucketRequestPayment (Maybe Text)

The base64-encoded 128-bit MD5 digest of the data. You must use this header as a message integrity check to verify that the request body was not corrupted in transit. For more information, see RFC 1864.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

putBucketRequestPayment_bucket :: Lens' PutBucketRequestPayment BucketName

The bucket name.

putBucketRequestPayment_requestPaymentConfiguration :: Lens' PutBucketRequestPayment RequestPaymentConfiguration

Container for Payer.

Destructuring the Response

data PutBucketRequestPaymentResponse

See: newPutBucketRequestPaymentResponse smart constructor.

Constructors

PutBucketRequestPaymentResponse'

newPutBucketRequestPaymentResponse :: PutBucketRequestPaymentResponse

Create a value of PutBucketRequestPaymentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.