Network.AWS.S3.PutBucketTagging

Description

Sets the tags for a bucket.

Use tags to organize your Amazon Web Services bill to reflect your own cost structure. To do this, sign up to get your Amazon Web Services account bill with tag key values included. Then, to see the cost of combined resources, organize your billing information according to resources with the same tag key values. For example, you can tag several resources with a specific application name, and then organize your billing information to see the total cost of that application across several services. For more information, see Cost Allocation and Tagging and Using Cost Allocation in Amazon S3 Bucket Tags.

When this operation sets the tags for a bucket, it will overwrite any current tags the bucket already has. You cannot use this operation to add tags to an existing list of tags.

To use this operation, you must have permissions to perform the s3:PutBucketTagging action. The bucket owner has this permission by default and can grant this permission to others. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to Your Amazon S3 Resources.

PutBucketTagging has the following special errors:

Error code: InvalidTagError

Description: The tag provided was not a valid tag. This error can occur if the tag did not pass input validation. For information about tag restrictions, see User-Defined Tag Restrictions and Amazon Web Services-Generated Cost Allocation Tag Restrictions.

Error code: MalformedXMLError

Description: The XML provided does not match the schema.

Error code: OperationAbortedError

Description: A conflicting conditional action is currently in progress against this resource. Please try again.

Error code: InternalError

Description: The service was unable to apply the provided tag to the bucket.

The following operations are related to PutBucketTagging:

GetBucketTagging

DeleteBucketTagging

Creating a Request

data PutBucketTagging

See: newPutBucketTagging smart constructor.

Constructors

PutBucketTagging'

Fields

contentMD5 :: Maybe Text

The base64-encoded 128-bit MD5 digest of the data. You must use this header as a message integrity check to verify that the request body was not corrupted in transit. For more information, see RFC 1864.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The bucket name.

tagging :: Tagging

Container for the TagSet and Tag elements.

newPutBucketTagging

Arguments

:: BucketName
$sel:bucket:PutBucketTagging'

-> Tagging
$sel:tagging:PutBucketTagging'

-> PutBucketTagging

Create a value of PutBucketTagging with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentMD5:PutBucketTagging', putBucketTagging_contentMD5 - The base64-encoded 128-bit MD5 digest of the data. You must use this header as a message integrity check to verify that the request body was not corrupted in transit. For more information, see RFC 1864.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

$sel:expectedBucketOwner:PutBucketTagging', putBucketTagging_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:PutBucketTagging', putBucketTagging_bucket - The bucket name.

$sel:tagging:PutBucketTagging', putBucketTagging_tagging - Container for the TagSet and Tag elements.

Request Lenses

putBucketTagging_contentMD5 :: Lens' PutBucketTagging (Maybe Text)

The base64-encoded 128-bit MD5 digest of the data. You must use this header as a message integrity check to verify that the request body was not corrupted in transit. For more information, see RFC 1864.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

putBucketTagging_expectedBucketOwner :: Lens' PutBucketTagging (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putBucketTagging_bucket :: Lens' PutBucketTagging BucketName

The bucket name.

putBucketTagging_tagging :: Lens' PutBucketTagging Tagging

Container for the TagSet and Tag elements.

Destructuring the Response

data PutBucketTaggingResponse

See: newPutBucketTaggingResponse smart constructor.

Constructors

PutBucketTaggingResponse'

newPutBucketTaggingResponse :: PutBucketTaggingResponse

Create a value of PutBucketTaggingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.