Network.AWS.S3.PutObjectAcl

Description

Uses the acl subresource to set the access control list (ACL) permissions for a new or existing object in an S3 bucket. You must have WRITE_ACP permission to set the ACL of an object. For more information, see What permissions can I grant? in the Amazon S3 User Guide.

This action is not supported by Amazon S3 on Outposts.

Depending on your application needs, you can choose to set the ACL on an object using either the request body or the headers. For example, if you have an existing application that updates a bucket ACL using the request body, you can continue to use that approach. For more information, see Access Control List (ACL) Overview in the Amazon S3 User Guide.

Access Permissions

You can set access permissions using one of the following methods:

Specify a canned ACL with the x-amz-acl request header. Amazon S3 supports a set of predefined ACLs, known as canned ACLs. Each canned ACL has a predefined set of grantees and permissions. Specify the canned ACL name as the value of x-amz-acl. If you use this header, you cannot use other access control-specific headers in your request. For more information, see Canned ACL.

Specify access permissions explicitly with the x-amz-grant-read, x-amz-grant-read-acp, x-amz-grant-write-acp, and x-amz-grant-full-control headers. When using these headers, you specify explicit access permissions and grantees (Amazon Web Services accounts or Amazon S3 groups) who will receive the permission. If you use these ACL-specific headers, you cannot use x-amz-acl header to set a canned ACL. These parameters map to the set of permissions that Amazon S3 supports in an ACL. For more information, see Access Control List (ACL) Overview.

You specify each grantee as a type=value pair, where the type is one of the following:

id – if the value specified is the canonical user ID of an Amazon Web Services account

uri – if you are granting permissions to a predefined group

emailAddress – if the value specified is the email address of an Amazon Web Services account

Using email addresses to specify a grantee is only supported in the following Amazon Web Services Regions:

US East (N. Virginia)

US West (N. California)

US West (Oregon)

Asia Pacific (Singapore)

Asia Pacific (Sydney)

Asia Pacific (Tokyo)

Europe (Ireland)

South America (São Paulo)

For a list of all the Amazon S3 supported Regions and endpoints, see Regions and Endpoints in the Amazon Web Services General Reference.

For example, the following x-amz-grant-read header grants list objects permission to the two Amazon Web Services accounts identified by their email addresses.

x-amz-grant-read: emailAddress="xyz@amazon.com", emailAddress="abc@amazon.com"

You can use either a canned ACL or specify access permissions explicitly. You cannot do both.

Grantee Values

You can specify the person (grantee) to whom you're assigning access rights (using request elements) in the following ways:

By the person's ID:

<Grantee xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:type="CanonicalUser"><ID><>ID<></ID><DisplayName><>GranteesEmail<></DisplayName> </Grantee>

DisplayName is optional and ignored in the request.

By URI:

<Grantee xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:type="Group"><URI><>http://acs.amazonaws.com/groups/global/AuthenticatedUsers<></URI></Grantee>

By Email address:

<Grantee xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:type="AmazonCustomerByEmail"><EmailAddress><>Grantees@email.com<></EmailAddress>lt;/Grantee>

The grantee is resolved to the CanonicalUser and, in a response to a GET Object acl request, appears as the CanonicalUser.

Using email addresses to specify a grantee is only supported in the following Amazon Web Services Regions:

US East (N. Virginia)

US West (N. California)

US West (Oregon)

Asia Pacific (Singapore)

Asia Pacific (Sydney)

Asia Pacific (Tokyo)

Europe (Ireland)

South America (São Paulo)

For a list of all the Amazon S3 supported Regions and endpoints, see Regions and Endpoints in the Amazon Web Services General Reference.

Versioning

The ACL of an object is set at the object version level. By default, PUT sets the ACL of the current version of an object. To set the ACL of a different version, use the versionId subresource.

Related Resources

CopyObject

GetObject

Creating a Request

data PutObjectAcl

See: newPutObjectAcl smart constructor.

Constructors

PutObjectAcl'

Fields

grantRead :: Maybe Text

Allows grantee to list the objects in the bucket.

This action is not supported by Amazon S3 on Outposts.

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

accessControlPolicy :: Maybe AccessControlPolicy

Contains the elements that set the ACL permissions for an object per grantee.

contentMD5 :: Maybe Text

The base64-encoded 128-bit MD5 digest of the data. This header must be used as a message integrity check to verify that the request body was not corrupted in transit. For more information, go to RFC 1864.>

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

versionId :: Maybe ObjectVersionId

VersionId used to reference a specific version of the object.

grantWriteACP :: Maybe Text

Allows grantee to write the ACL for the applicable bucket.

This action is not supported by Amazon S3 on Outposts.

acl :: Maybe ObjectCannedACL

The canned ACL to apply to the object. For more information, see Canned ACL.

grantReadACP :: Maybe Text

Allows grantee to read the bucket ACL.

This action is not supported by Amazon S3 on Outposts.

requestPayer :: Maybe RequestPayer

grantWrite :: Maybe Text

Allows grantee to create new objects in the bucket.

For the bucket and object owners of existing objects, also allows deletions and overwrites of those objects.

grantFullControl :: Maybe Text

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

This action is not supported by Amazon S3 on Outposts.

bucket :: BucketName

The bucket name that contains the object to which you want to attach the ACL.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

key :: ObjectKey

Key for which the PUT action was initiated.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

newPutObjectAcl

Arguments

:: BucketName
$sel:bucket:PutObjectAcl'

-> ObjectKey
$sel:key:PutObjectAcl'

-> PutObjectAcl

Create a value of PutObjectAcl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:grantRead:PutObjectAcl', putObjectAcl_grantRead - Allows grantee to list the objects in the bucket.

This action is not supported by Amazon S3 on Outposts.

$sel:expectedBucketOwner:PutObjectAcl', putObjectAcl_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:accessControlPolicy:PutObjectAcl', putObjectAcl_accessControlPolicy - Contains the elements that set the ACL permissions for an object per grantee.

$sel:contentMD5:PutObjectAcl', putObjectAcl_contentMD5 - The base64-encoded 128-bit MD5 digest of the data. This header must be used as a message integrity check to verify that the request body was not corrupted in transit. For more information, go to RFC 1864.>

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

$sel:versionId:PutObjectAcl', putObjectAcl_versionId - VersionId used to reference a specific version of the object.

$sel:grantWriteACP:PutObjectAcl', putObjectAcl_grantWriteACP - Allows grantee to write the ACL for the applicable bucket.

This action is not supported by Amazon S3 on Outposts.

$sel:acl:PutObjectAcl', putObjectAcl_acl - The canned ACL to apply to the object. For more information, see Canned ACL.

$sel:grantReadACP:PutObjectAcl', putObjectAcl_grantReadACP - Allows grantee to read the bucket ACL.

This action is not supported by Amazon S3 on Outposts.

$sel:requestPayer:PutObjectAcl', putObjectAcl_requestPayer - Undocumented member.

$sel:grantWrite:PutObjectAcl', putObjectAcl_grantWrite - Allows grantee to create new objects in the bucket.

For the bucket and object owners of existing objects, also allows deletions and overwrites of those objects.

$sel:grantFullControl:PutObjectAcl', putObjectAcl_grantFullControl - Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

This action is not supported by Amazon S3 on Outposts.

$sel:bucket:PutObjectAcl', putObjectAcl_bucket - The bucket name that contains the object to which you want to attach the ACL.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

$sel:key:PutObjectAcl', putObjectAcl_key - Key for which the PUT action was initiated.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

Request Lenses

putObjectAcl_grantRead :: Lens' PutObjectAcl (Maybe Text)

Allows grantee to list the objects in the bucket.

This action is not supported by Amazon S3 on Outposts.

putObjectAcl_expectedBucketOwner :: Lens' PutObjectAcl (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putObjectAcl_accessControlPolicy :: Lens' PutObjectAcl (Maybe AccessControlPolicy)

Contains the elements that set the ACL permissions for an object per grantee.

putObjectAcl_contentMD5 :: Lens' PutObjectAcl (Maybe Text)

The base64-encoded 128-bit MD5 digest of the data. This header must be used as a message integrity check to verify that the request body was not corrupted in transit. For more information, go to RFC 1864.>

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

putObjectAcl_versionId :: Lens' PutObjectAcl (Maybe ObjectVersionId)

VersionId used to reference a specific version of the object.

putObjectAcl_grantWriteACP :: Lens' PutObjectAcl (Maybe Text)

Allows grantee to write the ACL for the applicable bucket.

This action is not supported by Amazon S3 on Outposts.

putObjectAcl_acl :: Lens' PutObjectAcl (Maybe ObjectCannedACL)

The canned ACL to apply to the object. For more information, see Canned ACL.

putObjectAcl_grantReadACP :: Lens' PutObjectAcl (Maybe Text)

Allows grantee to read the bucket ACL.

This action is not supported by Amazon S3 on Outposts.

putObjectAcl_requestPayer :: Lens' PutObjectAcl (Maybe RequestPayer)

Undocumented member.

putObjectAcl_grantWrite :: Lens' PutObjectAcl (Maybe Text)

Allows grantee to create new objects in the bucket.

For the bucket and object owners of existing objects, also allows deletions and overwrites of those objects.

putObjectAcl_grantFullControl :: Lens' PutObjectAcl (Maybe Text)

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

This action is not supported by Amazon S3 on Outposts.

putObjectAcl_bucket :: Lens' PutObjectAcl BucketName

The bucket name that contains the object to which you want to attach the ACL.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

putObjectAcl_key :: Lens' PutObjectAcl ObjectKey

Key for which the PUT action was initiated.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

Destructuring the Response

data PutObjectAclResponse

See: newPutObjectAclResponse smart constructor.

Constructors

PutObjectAclResponse'

Fields

requestCharged :: Maybe RequestCharged

httpStatus :: Int

The response's http status code.

newPutObjectAclResponse

Arguments

:: Int
$sel:httpStatus:PutObjectAclResponse'

-> PutObjectAclResponse

Create a value of PutObjectAclResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestCharged:PutObjectAclResponse', putObjectAclResponse_requestCharged - Undocumented member.

$sel:httpStatus:PutObjectAclResponse', putObjectAclResponse_httpStatus - The response's http status code.

Response Lenses

putObjectAclResponse_requestCharged :: Lens' PutObjectAclResponse (Maybe RequestCharged)

Undocumented member.

putObjectAclResponse_httpStatus :: Lens' PutObjectAclResponse Int

The response's http status code.