Network.AWS.S3.PutObjectLegalHold

Description

Applies a Legal Hold configuration to the specified object. For more information, see Locking Objects.

This action is not supported by Amazon S3 on Outposts.

Creating a Request

data PutObjectLegalHold

See: newPutObjectLegalHold smart constructor.

Constructors

PutObjectLegalHold'

Fields

legalHold :: Maybe ObjectLockLegalHold

Container element for the Legal Hold configuration you want to apply to the specified object.

versionId :: Maybe ObjectVersionId

The version ID of the object that you want to place a Legal Hold on.

requestPayer :: Maybe RequestPayer

contentMD5 :: Maybe Text

The MD5 hash for the request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The bucket name containing the object that you want to place a Legal Hold on.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

key :: ObjectKey

The key name for the object that you want to place a Legal Hold on.

newPutObjectLegalHold

Arguments

:: BucketName
$sel:bucket:PutObjectLegalHold'

-> ObjectKey
$sel:key:PutObjectLegalHold'

-> PutObjectLegalHold

Create a value of PutObjectLegalHold with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:legalHold:PutObjectLegalHold', putObjectLegalHold_legalHold - Container element for the Legal Hold configuration you want to apply to the specified object.

$sel:versionId:PutObjectLegalHold', putObjectLegalHold_versionId - The version ID of the object that you want to place a Legal Hold on.

$sel:requestPayer:PutObjectLegalHold', putObjectLegalHold_requestPayer - Undocumented member.

$sel:contentMD5:PutObjectLegalHold', putObjectLegalHold_contentMD5 - The MD5 hash for the request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

$sel:expectedBucketOwner:PutObjectLegalHold', putObjectLegalHold_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:PutObjectLegalHold', putObjectLegalHold_bucket - The bucket name containing the object that you want to place a Legal Hold on.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

$sel:key:PutObjectLegalHold', putObjectLegalHold_key - The key name for the object that you want to place a Legal Hold on.

Request Lenses

putObjectLegalHold_legalHold :: Lens' PutObjectLegalHold (Maybe ObjectLockLegalHold)

Container element for the Legal Hold configuration you want to apply to the specified object.

putObjectLegalHold_versionId :: Lens' PutObjectLegalHold (Maybe ObjectVersionId)

The version ID of the object that you want to place a Legal Hold on.

putObjectLegalHold_requestPayer :: Lens' PutObjectLegalHold (Maybe RequestPayer)

Undocumented member.

putObjectLegalHold_contentMD5 :: Lens' PutObjectLegalHold (Maybe Text)

The MD5 hash for the request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

putObjectLegalHold_expectedBucketOwner :: Lens' PutObjectLegalHold (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putObjectLegalHold_bucket :: Lens' PutObjectLegalHold BucketName

The bucket name containing the object that you want to place a Legal Hold on.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

putObjectLegalHold_key :: Lens' PutObjectLegalHold ObjectKey

The key name for the object that you want to place a Legal Hold on.

Destructuring the Response

data PutObjectLegalHoldResponse

See: newPutObjectLegalHoldResponse smart constructor.

Constructors

PutObjectLegalHoldResponse'

Fields

requestCharged :: Maybe RequestCharged

httpStatus :: Int

The response's http status code.

newPutObjectLegalHoldResponse

Arguments

:: Int
$sel:httpStatus:PutObjectLegalHoldResponse'

-> PutObjectLegalHoldResponse

Create a value of PutObjectLegalHoldResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestCharged:PutObjectLegalHoldResponse', putObjectLegalHoldResponse_requestCharged - Undocumented member.

$sel:httpStatus:PutObjectLegalHoldResponse', putObjectLegalHoldResponse_httpStatus - The response's http status code.

Response Lenses

putObjectLegalHoldResponse_requestCharged :: Lens' PutObjectLegalHoldResponse (Maybe RequestCharged)

Undocumented member.

putObjectLegalHoldResponse_httpStatus :: Lens' PutObjectLegalHoldResponse Int

The response's http status code.