Network.AWS.S3.PutObjectRetention

Description

Places an Object Retention configuration on an object. For more information, see Locking Objects. Users or accounts require the s3:PutObjectRetention permission in order to place an Object Retention configuration on objects. Bypassing a Governance Retention configuration requires the s3:BypassGovernanceRetention permission.

This action is not supported by Amazon S3 on Outposts.

Permissions

When the Object Lock retention mode is set to compliance, you need s3:PutObjectRetention and s3:BypassGovernanceRetention permissions. For other requests to PutObjectRetention, only s3:PutObjectRetention permissions are required.

Creating a Request

data PutObjectRetention

See: newPutObjectRetention smart constructor.

Constructors

PutObjectRetention'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bypassGovernanceRetention :: Maybe Bool

Indicates whether this action should bypass Governance-mode restrictions.

contentMD5 :: Maybe Text

The MD5 hash for the request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

versionId :: Maybe ObjectVersionId

The version ID for the object that you want to apply this Object Retention configuration to.

retention :: Maybe ObjectLockRetention

The container element for the Object Retention configuration.

requestPayer :: Maybe RequestPayer

bucket :: BucketName

The bucket name that contains the object you want to apply this Object Retention configuration to.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

key :: ObjectKey

The key name for the object that you want to apply this Object Retention configuration to.

newPutObjectRetention

Arguments

:: BucketName
$sel:bucket:PutObjectRetention'

-> ObjectKey
$sel:key:PutObjectRetention'

-> PutObjectRetention

Create a value of PutObjectRetention with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:PutObjectRetention', putObjectRetention_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bypassGovernanceRetention:PutObjectRetention', putObjectRetention_bypassGovernanceRetention - Indicates whether this action should bypass Governance-mode restrictions.

$sel:contentMD5:PutObjectRetention', putObjectRetention_contentMD5 - The MD5 hash for the request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

$sel:versionId:PutObjectRetention', putObjectRetention_versionId - The version ID for the object that you want to apply this Object Retention configuration to.

$sel:retention:PutObjectRetention', putObjectRetention_retention - The container element for the Object Retention configuration.

$sel:requestPayer:PutObjectRetention', putObjectRetention_requestPayer - Undocumented member.

$sel:bucket:PutObjectRetention', putObjectRetention_bucket - The bucket name that contains the object you want to apply this Object Retention configuration to.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

$sel:key:PutObjectRetention', putObjectRetention_key - The key name for the object that you want to apply this Object Retention configuration to.

Request Lenses

putObjectRetention_expectedBucketOwner :: Lens' PutObjectRetention (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putObjectRetention_bypassGovernanceRetention :: Lens' PutObjectRetention (Maybe Bool)

Indicates whether this action should bypass Governance-mode restrictions.

putObjectRetention_contentMD5 :: Lens' PutObjectRetention (Maybe Text)

The MD5 hash for the request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

putObjectRetention_versionId :: Lens' PutObjectRetention (Maybe ObjectVersionId)

The version ID for the object that you want to apply this Object Retention configuration to.

putObjectRetention_retention :: Lens' PutObjectRetention (Maybe ObjectLockRetention)

The container element for the Object Retention configuration.

putObjectRetention_requestPayer :: Lens' PutObjectRetention (Maybe RequestPayer)

Undocumented member.

putObjectRetention_bucket :: Lens' PutObjectRetention BucketName

The bucket name that contains the object you want to apply this Object Retention configuration to.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

putObjectRetention_key :: Lens' PutObjectRetention ObjectKey

The key name for the object that you want to apply this Object Retention configuration to.

Destructuring the Response

data PutObjectRetentionResponse

See: newPutObjectRetentionResponse smart constructor.

Constructors

PutObjectRetentionResponse'

Fields

requestCharged :: Maybe RequestCharged

httpStatus :: Int

The response's http status code.

newPutObjectRetentionResponse

Arguments

:: Int
$sel:httpStatus:PutObjectRetentionResponse'

-> PutObjectRetentionResponse

Create a value of PutObjectRetentionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestCharged:PutObjectRetentionResponse', putObjectRetentionResponse_requestCharged - Undocumented member.

$sel:httpStatus:PutObjectRetentionResponse', putObjectRetentionResponse_httpStatus - The response's http status code.

Response Lenses

putObjectRetentionResponse_requestCharged :: Lens' PutObjectRetentionResponse (Maybe RequestCharged)

Undocumented member.

putObjectRetentionResponse_httpStatus :: Lens' PutObjectRetentionResponse Int

The response's http status code.