Network.AWS.S3.Types.AnalyticsAndOperator

Description

Documentation

data AnalyticsAndOperator

A conjunction (logical AND) of predicates, which is used in evaluating a metrics filter. The operator must have at least two predicates in any combination, and an object must match all of the predicates for the filter to apply.

See: newAnalyticsAndOperator smart constructor.

Constructors

AnalyticsAndOperator'

Fields

prefix :: Maybe Text

The prefix to use when evaluating an AND predicate: The prefix that an object must have to be included in the metrics results.

tags :: Maybe [Tag]

The list of tags to use when evaluating an AND predicate.

newAnalyticsAndOperator :: AnalyticsAndOperator

Create a value of AnalyticsAndOperator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:AnalyticsAndOperator', analyticsAndOperator_prefix - The prefix to use when evaluating an AND predicate: The prefix that an object must have to be included in the metrics results.

$sel:tags:AnalyticsAndOperator', analyticsAndOperator_tags - The list of tags to use when evaluating an AND predicate.

analyticsAndOperator_prefix :: Lens' AnalyticsAndOperator (Maybe Text)

The prefix to use when evaluating an AND predicate: The prefix that an object must have to be included in the metrics results.

analyticsAndOperator_tags :: Lens' AnalyticsAndOperator (Maybe [Tag])

The list of tags to use when evaluating an AND predicate.