Network.AWS.S3.Types.AnalyticsConfiguration

Description

Documentation

data AnalyticsConfiguration

Specifies the configuration and any analyses for the analytics filter of an Amazon S3 bucket.

See: newAnalyticsConfiguration smart constructor.

Constructors

AnalyticsConfiguration'

Fields

filter' :: Maybe AnalyticsFilter

The filter used to describe a set of objects for analyses. A filter must have exactly one prefix, one tag, or one conjunction (AnalyticsAndOperator). If no filter is provided, all objects will be considered in any analysis.

id :: Text

The ID that identifies the analytics configuration.

storageClassAnalysis :: StorageClassAnalysis

Contains data related to access patterns to be collected and made available to analyze the tradeoffs between different storage classes.

newAnalyticsConfiguration

Arguments

:: Text
$sel:id:AnalyticsConfiguration'

-> StorageClassAnalysis
$sel:storageClassAnalysis:AnalyticsConfiguration'

-> AnalyticsConfiguration

Create a value of AnalyticsConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filter':AnalyticsConfiguration', analyticsConfiguration_filter - The filter used to describe a set of objects for analyses. A filter must have exactly one prefix, one tag, or one conjunction (AnalyticsAndOperator). If no filter is provided, all objects will be considered in any analysis.

$sel:id:AnalyticsConfiguration', analyticsConfiguration_id - The ID that identifies the analytics configuration.

$sel:storageClassAnalysis:AnalyticsConfiguration', analyticsConfiguration_storageClassAnalysis - Contains data related to access patterns to be collected and made available to analyze the tradeoffs between different storage classes.

analyticsConfiguration_filter :: Lens' AnalyticsConfiguration (Maybe AnalyticsFilter)

The filter used to describe a set of objects for analyses. A filter must have exactly one prefix, one tag, or one conjunction (AnalyticsAndOperator). If no filter is provided, all objects will be considered in any analysis.

analyticsConfiguration_id :: Lens' AnalyticsConfiguration Text

The ID that identifies the analytics configuration.

analyticsConfiguration_storageClassAnalysis :: Lens' AnalyticsConfiguration StorageClassAnalysis

Contains data related to access patterns to be collected and made available to analyze the tradeoffs between different storage classes.