Network.AWS.S3.Types.AnalyticsS3BucketDestination

Description

Documentation

data AnalyticsS3BucketDestination

Contains information about where to publish the analytics results.

See: newAnalyticsS3BucketDestination smart constructor.

Constructors

AnalyticsS3BucketDestination'

Fields

prefix :: Maybe Text

The prefix to use when exporting data. The prefix is prepended to all results.

bucketAccountId :: Maybe Text

The account ID that owns the destination S3 bucket. If no account ID is provided, the owner is not validated before exporting data.

Although this value is optional, we strongly recommend that you set it to help prevent problems if the destination bucket ownership changes.

format :: AnalyticsS3ExportFileFormat

Specifies the file format used when exporting data to Amazon S3.

bucket :: BucketName

The Amazon Resource Name (ARN) of the bucket to which data is exported.

newAnalyticsS3BucketDestination

Arguments

:: AnalyticsS3ExportFileFormat
$sel:format:AnalyticsS3BucketDestination'

-> BucketName
$sel:bucket:AnalyticsS3BucketDestination'

-> AnalyticsS3BucketDestination

Create a value of AnalyticsS3BucketDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:AnalyticsS3BucketDestination', analyticsS3BucketDestination_prefix - The prefix to use when exporting data. The prefix is prepended to all results.

$sel:bucketAccountId:AnalyticsS3BucketDestination', analyticsS3BucketDestination_bucketAccountId - The account ID that owns the destination S3 bucket. If no account ID is provided, the owner is not validated before exporting data.

Although this value is optional, we strongly recommend that you set it to help prevent problems if the destination bucket ownership changes.

$sel:format:AnalyticsS3BucketDestination', analyticsS3BucketDestination_format - Specifies the file format used when exporting data to Amazon S3.

$sel:bucket:AnalyticsS3BucketDestination', analyticsS3BucketDestination_bucket - The Amazon Resource Name (ARN) of the bucket to which data is exported.

analyticsS3BucketDestination_prefix :: Lens' AnalyticsS3BucketDestination (Maybe Text)

The prefix to use when exporting data. The prefix is prepended to all results.

analyticsS3BucketDestination_bucketAccountId :: Lens' AnalyticsS3BucketDestination (Maybe Text)

The account ID that owns the destination S3 bucket. If no account ID is provided, the owner is not validated before exporting data.

Although this value is optional, we strongly recommend that you set it to help prevent problems if the destination bucket ownership changes.

analyticsS3BucketDestination_format :: Lens' AnalyticsS3BucketDestination AnalyticsS3ExportFileFormat

Specifies the file format used when exporting data to Amazon S3.

analyticsS3BucketDestination_bucket :: Lens' AnalyticsS3BucketDestination BucketName

The Amazon Resource Name (ARN) of the bucket to which data is exported.