Network.AWS.S3.Types.CSVInput

Description

Documentation

data CSVInput

Describes how an uncompressed comma-separated values (CSV)-formatted input object is formatted.

See: newCSVInput smart constructor.

Constructors

CSVInput'

Fields

allowQuotedRecordDelimiter :: Maybe Bool

Specifies that CSV field values may contain quoted record delimiters and such records should be allowed. Default value is FALSE. Setting this value to TRUE may lower performance.

recordDelimiter :: Maybe Text

A single character used to separate individual records in the input. Instead of the default value, you can specify an arbitrary delimiter.

quoteCharacter :: Maybe Text

A single character used for escaping when the field delimiter is part of the value. For example, if the value is a, b, Amazon S3 wraps this field value in quotation marks, as follows: " a , b ".

Type: String

Default: "

Ancestors: CSV

fileHeaderInfo :: Maybe FileHeaderInfo

Describes the first line of input. Valid values are:

NONE: First line is not a header.

IGNORE: First line is a header, but you can't use the header values to indicate the column in an expression. You can use column position (such as _1, _2, …) to indicate the column (SELECT s._1 FROM OBJECT s).

Use: First line is a header, and you can use the header value to identify a column in an expression (SELECT "name" FROM OBJECT).

fieldDelimiter :: Maybe Text

A single character used to separate individual fields in a record. You can specify an arbitrary delimiter.

comments :: Maybe Text

A single character used to indicate that a row should be ignored when the character is present at the start of that row. You can specify any character to indicate a comment line.

quoteEscapeCharacter :: Maybe Text

A single character used for escaping the quotation mark character inside an already escaped value. For example, the value """ a , b """ is parsed as " a , b ".

newCSVInput :: CSVInput

Create a value of CSVInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:allowQuotedRecordDelimiter:CSVInput', cSVInput_allowQuotedRecordDelimiter - Specifies that CSV field values may contain quoted record delimiters and such records should be allowed. Default value is FALSE. Setting this value to TRUE may lower performance.

$sel:recordDelimiter:CSVInput', cSVInput_recordDelimiter - A single character used to separate individual records in the input. Instead of the default value, you can specify an arbitrary delimiter.

$sel:quoteCharacter:CSVInput', cSVInput_quoteCharacter - A single character used for escaping when the field delimiter is part of the value. For example, if the value is a, b, Amazon S3 wraps this field value in quotation marks, as follows: " a , b ".

Type: String

Default: "

Ancestors: CSV

$sel:fileHeaderInfo:CSVInput', cSVInput_fileHeaderInfo - Describes the first line of input. Valid values are:

NONE: First line is not a header.

IGNORE: First line is a header, but you can't use the header values to indicate the column in an expression. You can use column position (such as _1, _2, …) to indicate the column (SELECT s._1 FROM OBJECT s).

Use: First line is a header, and you can use the header value to identify a column in an expression (SELECT "name" FROM OBJECT).

$sel:fieldDelimiter:CSVInput', cSVInput_fieldDelimiter - A single character used to separate individual fields in a record. You can specify an arbitrary delimiter.

$sel:comments:CSVInput', cSVInput_comments - A single character used to indicate that a row should be ignored when the character is present at the start of that row. You can specify any character to indicate a comment line.

$sel:quoteEscapeCharacter:CSVInput', cSVInput_quoteEscapeCharacter - A single character used for escaping the quotation mark character inside an already escaped value. For example, the value """ a , b """ is parsed as " a , b ".

cSVInput_allowQuotedRecordDelimiter :: Lens' CSVInput (Maybe Bool)

Specifies that CSV field values may contain quoted record delimiters and such records should be allowed. Default value is FALSE. Setting this value to TRUE may lower performance.

cSVInput_recordDelimiter :: Lens' CSVInput (Maybe Text)

A single character used to separate individual records in the input. Instead of the default value, you can specify an arbitrary delimiter.

cSVInput_quoteCharacter :: Lens' CSVInput (Maybe Text)

A single character used for escaping when the field delimiter is part of the value. For example, if the value is a, b, Amazon S3 wraps this field value in quotation marks, as follows: " a , b ".

Type: String

Default: "

Ancestors: CSV

cSVInput_fileHeaderInfo :: Lens' CSVInput (Maybe FileHeaderInfo)

Describes the first line of input. Valid values are:

NONE: First line is not a header.

IGNORE: First line is a header, but you can't use the header values to indicate the column in an expression. You can use column position (such as _1, _2, …) to indicate the column (SELECT s._1 FROM OBJECT s).

Use: First line is a header, and you can use the header value to identify a column in an expression (SELECT "name" FROM OBJECT).

cSVInput_fieldDelimiter :: Lens' CSVInput (Maybe Text)

A single character used to separate individual fields in a record. You can specify an arbitrary delimiter.

cSVInput_comments :: Lens' CSVInput (Maybe Text)

A single character used to indicate that a row should be ignored when the character is present at the start of that row. You can specify any character to indicate a comment line.

cSVInput_quoteEscapeCharacter :: Lens' CSVInput (Maybe Text)

A single character used for escaping the quotation mark character inside an already escaped value. For example, the value """ a , b """ is parsed as " a , b ".