Network.AWS.S3.Types.DefaultRetention

Description

Documentation

data DefaultRetention

The container element for specifying the default Object Lock retention settings for new objects placed in the specified bucket.

The DefaultRetention settings require both a mode and a period.

The DefaultRetention period can be either Days or Years but you must select one. You cannot specify Days and Years at the same time.

See: newDefaultRetention smart constructor.

Constructors

DefaultRetention'

Fields

days :: Maybe Int

The number of days that you want to specify for the default retention period. Must be used with Mode.

mode :: Maybe ObjectLockRetentionMode

The default Object Lock retention mode you want to apply to new objects placed in the specified bucket. Must be used with either Days or Years.

years :: Maybe Int

The number of years that you want to specify for the default retention period. Must be used with Mode.

newDefaultRetention :: DefaultRetention

Create a value of DefaultRetention with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:days:DefaultRetention', defaultRetention_days - The number of days that you want to specify for the default retention period. Must be used with Mode.

$sel:mode:DefaultRetention', defaultRetention_mode - The default Object Lock retention mode you want to apply to new objects placed in the specified bucket. Must be used with either Days or Years.

$sel:years:DefaultRetention', defaultRetention_years - The number of years that you want to specify for the default retention period. Must be used with Mode.

defaultRetention_days :: Lens' DefaultRetention (Maybe Int)

The number of days that you want to specify for the default retention period. Must be used with Mode.

defaultRetention_mode :: Lens' DefaultRetention (Maybe ObjectLockRetentionMode)

The default Object Lock retention mode you want to apply to new objects placed in the specified bucket. Must be used with either Days or Years.

defaultRetention_years :: Lens' DefaultRetention (Maybe Int)

The number of years that you want to specify for the default retention period. Must be used with Mode.