Network.AWS.S3.Types.DeleteMarkerReplication

Description

Documentation

data DeleteMarkerReplication

Specifies whether Amazon S3 replicates delete markers. If you specify a Filter in your replication configuration, you must also include a DeleteMarkerReplication element. If your Filter includes a Tag element, the DeleteMarkerReplication Status must be set to Disabled, because Amazon S3 does not support replicating delete markers for tag-based rules. For an example configuration, see Basic Rule Configuration.

For more information about delete marker replication, see Basic Rule Configuration.

If you are using an earlier version of the replication configuration, Amazon S3 handles replication of delete markers differently. For more information, see Backward Compatibility.

See: newDeleteMarkerReplication smart constructor.

Constructors

DeleteMarkerReplication'

Fields

status :: Maybe DeleteMarkerReplicationStatus

Indicates whether to replicate delete markers.

Indicates whether to replicate delete markers.

newDeleteMarkerReplication :: DeleteMarkerReplication

Create a value of DeleteMarkerReplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DeleteMarkerReplication', deleteMarkerReplication_status - Indicates whether to replicate delete markers.

Indicates whether to replicate delete markers.

deleteMarkerReplication_status :: Lens' DeleteMarkerReplication (Maybe DeleteMarkerReplicationStatus)

Indicates whether to replicate delete markers.

Indicates whether to replicate delete markers.