Network.AWS.S3.Types.DeletedObject

Description

Documentation

data DeletedObject

Information about the deleted object.

See: newDeletedObject smart constructor.

Constructors

DeletedObject'

Fields

key :: Maybe ObjectKey

The name of the deleted object.

deleteMarkerVersionId :: Maybe Text

The version ID of the delete marker created as a result of the DELETE operation. If you delete a specific object version, the value returned by this header is the version ID of the object version deleted.

deleteMarker :: Maybe Bool

Specifies whether the versioned object that was permanently deleted was (true) or was not (false) a delete marker. In a simple DELETE, this header indicates whether (true) or not (false) a delete marker was created.

versionId :: Maybe ObjectVersionId

The version ID of the deleted object.

newDeletedObject :: DeletedObject

Create a value of DeletedObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:DeletedObject', deletedObject_key - The name of the deleted object.

$sel:deleteMarkerVersionId:DeletedObject', deletedObject_deleteMarkerVersionId - The version ID of the delete marker created as a result of the DELETE operation. If you delete a specific object version, the value returned by this header is the version ID of the object version deleted.

$sel:deleteMarker:DeletedObject', deletedObject_deleteMarker - Specifies whether the versioned object that was permanently deleted was (true) or was not (false) a delete marker. In a simple DELETE, this header indicates whether (true) or not (false) a delete marker was created.

$sel:versionId:DeletedObject', deletedObject_versionId - The version ID of the deleted object.

deletedObject_key :: Lens' DeletedObject (Maybe ObjectKey)

The name of the deleted object.

deletedObject_deleteMarkerVersionId :: Lens' DeletedObject (Maybe Text)

The version ID of the delete marker created as a result of the DELETE operation. If you delete a specific object version, the value returned by this header is the version ID of the object version deleted.

deletedObject_deleteMarker :: Lens' DeletedObject (Maybe Bool)

Specifies whether the versioned object that was permanently deleted was (true) or was not (false) a delete marker. In a simple DELETE, this header indicates whether (true) or not (false) a delete marker was created.

deletedObject_versionId :: Lens' DeletedObject (Maybe ObjectVersionId)

The version ID of the deleted object.