Network.AWS.S3.Types.Destination

Description

Documentation

data Destination

Specifies information about where to publish analysis or configuration results for an Amazon S3 bucket and S3 Replication Time Control (S3 RTC).

See: newDestination smart constructor.

Constructors

Destination'

Fields

metrics :: Maybe Metrics

A container specifying replication metrics-related settings enabling replication metrics and events.

accessControlTranslation :: Maybe AccessControlTranslation

Specify this only in a cross-account scenario (where source and destination bucket owners are not the same), and you want to change replica ownership to the Amazon Web Services account that owns the destination bucket. If this is not specified in the replication configuration, the replicas are owned by same Amazon Web Services account that owns the source object.

account :: Maybe Text

Destination bucket owner account ID. In a cross-account scenario, if you direct Amazon S3 to change replica ownership to the Amazon Web Services account that owns the destination bucket by specifying the AccessControlTranslation property, this is the account ID of the destination bucket owner. For more information, see Replication Additional Configuration: Changing the Replica Owner in the Amazon S3 User Guide.

storageClass :: Maybe StorageClass

The storage class to use when replicating objects, such as S3 Standard or reduced redundancy. By default, Amazon S3 uses the storage class of the source object to create the object replica.

For valid values, see the StorageClass element of the PUT Bucket replication action in the Amazon S3 API Reference.

encryptionConfiguration :: Maybe EncryptionConfiguration

A container that provides information about encryption. If SourceSelectionCriteria is specified, you must specify this element.

replicationTime :: Maybe ReplicationTime

A container specifying S3 Replication Time Control (S3 RTC), including whether S3 RTC is enabled and the time when all objects and operations on objects must be replicated. Must be specified together with a Metrics block.

bucket :: BucketName

The Amazon Resource Name (ARN) of the bucket where you want Amazon S3 to store the results.

newDestination

Arguments

:: BucketName
$sel:bucket:Destination'

-> Destination

Create a value of Destination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metrics:Destination', destination_metrics - A container specifying replication metrics-related settings enabling replication metrics and events.

$sel:accessControlTranslation:Destination', destination_accessControlTranslation - Specify this only in a cross-account scenario (where source and destination bucket owners are not the same), and you want to change replica ownership to the Amazon Web Services account that owns the destination bucket. If this is not specified in the replication configuration, the replicas are owned by same Amazon Web Services account that owns the source object.

$sel:account:Destination', destination_account - Destination bucket owner account ID. In a cross-account scenario, if you direct Amazon S3 to change replica ownership to the Amazon Web Services account that owns the destination bucket by specifying the AccessControlTranslation property, this is the account ID of the destination bucket owner. For more information, see Replication Additional Configuration: Changing the Replica Owner in the Amazon S3 User Guide.

$sel:storageClass:Destination', destination_storageClass - The storage class to use when replicating objects, such as S3 Standard or reduced redundancy. By default, Amazon S3 uses the storage class of the source object to create the object replica.

For valid values, see the StorageClass element of the PUT Bucket replication action in the Amazon S3 API Reference.

$sel:encryptionConfiguration:Destination', destination_encryptionConfiguration - A container that provides information about encryption. If SourceSelectionCriteria is specified, you must specify this element.

$sel:replicationTime:Destination', destination_replicationTime - A container specifying S3 Replication Time Control (S3 RTC), including whether S3 RTC is enabled and the time when all objects and operations on objects must be replicated. Must be specified together with a Metrics block.

$sel:bucket:Destination', destination_bucket - The Amazon Resource Name (ARN) of the bucket where you want Amazon S3 to store the results.

destination_metrics :: Lens' Destination (Maybe Metrics)

A container specifying replication metrics-related settings enabling replication metrics and events.

destination_accessControlTranslation :: Lens' Destination (Maybe AccessControlTranslation)

Specify this only in a cross-account scenario (where source and destination bucket owners are not the same), and you want to change replica ownership to the Amazon Web Services account that owns the destination bucket. If this is not specified in the replication configuration, the replicas are owned by same Amazon Web Services account that owns the source object.

destination_account :: Lens' Destination (Maybe Text)

Destination bucket owner account ID. In a cross-account scenario, if you direct Amazon S3 to change replica ownership to the Amazon Web Services account that owns the destination bucket by specifying the AccessControlTranslation property, this is the account ID of the destination bucket owner. For more information, see Replication Additional Configuration: Changing the Replica Owner in the Amazon S3 User Guide.

destination_storageClass :: Lens' Destination (Maybe StorageClass)

The storage class to use when replicating objects, such as S3 Standard or reduced redundancy. By default, Amazon S3 uses the storage class of the source object to create the object replica.

For valid values, see the StorageClass element of the PUT Bucket replication action in the Amazon S3 API Reference.

destination_encryptionConfiguration :: Lens' Destination (Maybe EncryptionConfiguration)

A container that provides information about encryption. If SourceSelectionCriteria is specified, you must specify this element.

destination_replicationTime :: Lens' Destination (Maybe ReplicationTime)

A container specifying S3 Replication Time Control (S3 RTC), including whether S3 RTC is enabled and the time when all objects and operations on objects must be replicated. Must be specified together with a Metrics block.

destination_bucket :: Lens' Destination BucketName

The Amazon Resource Name (ARN) of the bucket where you want Amazon S3 to store the results.