Network.AWS.S3.Types.InventoryConfiguration

Description

Documentation

data InventoryConfiguration

Specifies the inventory configuration for an Amazon S3 bucket. For more information, see GET Bucket inventory in the Amazon S3 API Reference.

See: newInventoryConfiguration smart constructor.

Constructors

InventoryConfiguration'

Fields

optionalFields :: Maybe [InventoryOptionalField]

Contains the optional fields that are included in the inventory results.

filter' :: Maybe InventoryFilter

Specifies an inventory filter. The inventory only includes objects that meet the filter's criteria.

destination :: InventoryDestination

Contains information about where to publish the inventory results.

isEnabled :: Bool

Specifies whether the inventory is enabled or disabled. If set to True, an inventory list is generated. If set to False, no inventory list is generated.

id :: Text

The ID used to identify the inventory configuration.

includedObjectVersions :: InventoryIncludedObjectVersions

Object versions to include in the inventory list. If set to All, the list includes all the object versions, which adds the version-related fields VersionId, IsLatest, and DeleteMarker to the list. If set to Current, the list does not contain these version-related fields.

schedule :: InventorySchedule

Specifies the schedule for generating inventory results.

newInventoryConfiguration

Arguments

:: InventoryDestination
$sel:destination:InventoryConfiguration'

-> Bool
$sel:isEnabled:InventoryConfiguration'

-> Text
$sel:id:InventoryConfiguration'

-> InventoryIncludedObjectVersions
$sel:includedObjectVersions:InventoryConfiguration'

-> InventorySchedule
$sel:schedule:InventoryConfiguration'

-> InventoryConfiguration

Create a value of InventoryConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:optionalFields:InventoryConfiguration', inventoryConfiguration_optionalFields - Contains the optional fields that are included in the inventory results.

$sel:filter':InventoryConfiguration', inventoryConfiguration_filter - Specifies an inventory filter. The inventory only includes objects that meet the filter's criteria.

$sel:destination:InventoryConfiguration', inventoryConfiguration_destination - Contains information about where to publish the inventory results.

$sel:isEnabled:InventoryConfiguration', inventoryConfiguration_isEnabled - Specifies whether the inventory is enabled or disabled. If set to True, an inventory list is generated. If set to False, no inventory list is generated.

$sel:id:InventoryConfiguration', inventoryConfiguration_id - The ID used to identify the inventory configuration.

$sel:includedObjectVersions:InventoryConfiguration', inventoryConfiguration_includedObjectVersions - Object versions to include in the inventory list. If set to All, the list includes all the object versions, which adds the version-related fields VersionId, IsLatest, and DeleteMarker to the list. If set to Current, the list does not contain these version-related fields.

$sel:schedule:InventoryConfiguration', inventoryConfiguration_schedule - Specifies the schedule for generating inventory results.

inventoryConfiguration_optionalFields :: Lens' InventoryConfiguration (Maybe [InventoryOptionalField])

Contains the optional fields that are included in the inventory results.

inventoryConfiguration_filter :: Lens' InventoryConfiguration (Maybe InventoryFilter)

Specifies an inventory filter. The inventory only includes objects that meet the filter's criteria.

inventoryConfiguration_destination :: Lens' InventoryConfiguration InventoryDestination

Contains information about where to publish the inventory results.

inventoryConfiguration_isEnabled :: Lens' InventoryConfiguration Bool

Specifies whether the inventory is enabled or disabled. If set to True, an inventory list is generated. If set to False, no inventory list is generated.

inventoryConfiguration_id :: Lens' InventoryConfiguration Text

The ID used to identify the inventory configuration.

inventoryConfiguration_includedObjectVersions :: Lens' InventoryConfiguration InventoryIncludedObjectVersions

Object versions to include in the inventory list. If set to All, the list includes all the object versions, which adds the version-related fields VersionId, IsLatest, and DeleteMarker to the list. If set to Current, the list does not contain these version-related fields.

inventoryConfiguration_schedule :: Lens' InventoryConfiguration InventorySchedule

Specifies the schedule for generating inventory results.