Network.AWS.S3.Types.InventoryS3BucketDestination

Description

Documentation

data InventoryS3BucketDestination

Contains the bucket name, file format, bucket owner (optional), and prefix (optional) where inventory results are published.

See: newInventoryS3BucketDestination smart constructor.

Constructors

InventoryS3BucketDestination'

Fields

prefix :: Maybe Text

The prefix that is prepended to all inventory results.

accountId :: Maybe Text

The account ID that owns the destination S3 bucket. If no account ID is provided, the owner is not validated before exporting data.

Although this value is optional, we strongly recommend that you set it to help prevent problems if the destination bucket ownership changes.

encryption :: Maybe InventoryEncryption

Contains the type of server-side encryption used to encrypt the inventory results.

bucket :: BucketName

The Amazon Resource Name (ARN) of the bucket where inventory results will be published.

format :: InventoryFormat

Specifies the output format of the inventory results.

newInventoryS3BucketDestination

Arguments

:: BucketName
$sel:bucket:InventoryS3BucketDestination'

-> InventoryFormat
$sel:format:InventoryS3BucketDestination'

-> InventoryS3BucketDestination

Create a value of InventoryS3BucketDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:InventoryS3BucketDestination', inventoryS3BucketDestination_prefix - The prefix that is prepended to all inventory results.

$sel:accountId:InventoryS3BucketDestination', inventoryS3BucketDestination_accountId - The account ID that owns the destination S3 bucket. If no account ID is provided, the owner is not validated before exporting data.

Although this value is optional, we strongly recommend that you set it to help prevent problems if the destination bucket ownership changes.

$sel:encryption:InventoryS3BucketDestination', inventoryS3BucketDestination_encryption - Contains the type of server-side encryption used to encrypt the inventory results.

$sel:bucket:InventoryS3BucketDestination', inventoryS3BucketDestination_bucket - The Amazon Resource Name (ARN) of the bucket where inventory results will be published.

$sel:format:InventoryS3BucketDestination', inventoryS3BucketDestination_format - Specifies the output format of the inventory results.

inventoryS3BucketDestination_prefix :: Lens' InventoryS3BucketDestination (Maybe Text)

The prefix that is prepended to all inventory results.

inventoryS3BucketDestination_accountId :: Lens' InventoryS3BucketDestination (Maybe Text)

The account ID that owns the destination S3 bucket. If no account ID is provided, the owner is not validated before exporting data.

Although this value is optional, we strongly recommend that you set it to help prevent problems if the destination bucket ownership changes.

inventoryS3BucketDestination_encryption :: Lens' InventoryS3BucketDestination (Maybe InventoryEncryption)

Contains the type of server-side encryption used to encrypt the inventory results.

inventoryS3BucketDestination_bucket :: Lens' InventoryS3BucketDestination BucketName

The Amazon Resource Name (ARN) of the bucket where inventory results will be published.

inventoryS3BucketDestination_format :: Lens' InventoryS3BucketDestination InventoryFormat

Specifies the output format of the inventory results.