Network.AWS.S3.Types.LifecycleExpiration

Description

Documentation

data LifecycleExpiration

Container for the expiration for the lifecycle of the object.

See: newLifecycleExpiration smart constructor.

Constructors

LifecycleExpiration'

Fields

days :: Maybe Int

Indicates the lifetime, in days, of the objects that are subject to the rule. The value must be a non-zero positive integer.

expiredObjectDeleteMarker :: Maybe Bool

Indicates whether Amazon S3 will remove a delete marker with no noncurrent versions. If set to true, the delete marker will be expired; if set to false the policy takes no action. This cannot be specified with Days or Date in a Lifecycle Expiration Policy.

date :: Maybe ISO8601

Indicates at what date the object is to be moved or deleted. Should be in GMT ISO 8601 Format.

newLifecycleExpiration :: LifecycleExpiration

Create a value of LifecycleExpiration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:days:LifecycleExpiration', lifecycleExpiration_days - Indicates the lifetime, in days, of the objects that are subject to the rule. The value must be a non-zero positive integer.

$sel:expiredObjectDeleteMarker:LifecycleExpiration', lifecycleExpiration_expiredObjectDeleteMarker - Indicates whether Amazon S3 will remove a delete marker with no noncurrent versions. If set to true, the delete marker will be expired; if set to false the policy takes no action. This cannot be specified with Days or Date in a Lifecycle Expiration Policy.

$sel:date:LifecycleExpiration', lifecycleExpiration_date - Indicates at what date the object is to be moved or deleted. Should be in GMT ISO 8601 Format.

lifecycleExpiration_days :: Lens' LifecycleExpiration (Maybe Int)

Indicates the lifetime, in days, of the objects that are subject to the rule. The value must be a non-zero positive integer.

lifecycleExpiration_expiredObjectDeleteMarker :: Lens' LifecycleExpiration (Maybe Bool)

Indicates whether Amazon S3 will remove a delete marker with no noncurrent versions. If set to true, the delete marker will be expired; if set to false the policy takes no action. This cannot be specified with Days or Date in a Lifecycle Expiration Policy.

lifecycleExpiration_date :: Lens' LifecycleExpiration (Maybe UTCTime)

Indicates at what date the object is to be moved or deleted. Should be in GMT ISO 8601 Format.