Network.AWS.S3.Types.LifecycleRule

Description

Documentation

data LifecycleRule

A lifecycle rule for individual objects in an Amazon S3 bucket.

See: newLifecycleRule smart constructor.

Constructors

LifecycleRule'

Fields

transitions :: Maybe [Transition]

Specifies when an Amazon S3 object transitions to a specified storage class.

noncurrentVersionExpiration :: Maybe NoncurrentVersionExpiration

prefix :: Maybe Text

Prefix identifying one or more objects to which the rule applies. This is no longer used; use Filter instead.

Replacement must be made for object keys containing special characters (such as carriage returns) when using XML requests. For more information, see XML related object key constraints.

noncurrentVersionTransitions :: Maybe [NoncurrentVersionTransition]

Specifies the transition rule for the lifecycle rule that describes when noncurrent objects transition to a specific storage class. If your bucket is versioning-enabled (or versioning is suspended), you can set this action to request that Amazon S3 transition noncurrent object versions to a specific storage class at a set period in the object's lifetime.

expiration :: Maybe LifecycleExpiration

Specifies the expiration for the lifecycle of the object in the form of date, days and, whether the object has a delete marker.

id :: Maybe Text

Unique identifier for the rule. The value cannot be longer than 255 characters.

filter' :: Maybe LifecycleRuleFilter

The Filter is used to identify objects that a Lifecycle Rule applies to. A Filter must have exactly one of Prefix, Tag, or And specified. Filter is required if the LifecycleRule does not containt a Prefix element.

abortIncompleteMultipartUpload :: Maybe AbortIncompleteMultipartUpload

status :: ExpirationStatus

If 'Enabled', the rule is currently being applied. If 'Disabled', the rule is not currently being applied.

newLifecycleRule

Arguments

:: ExpirationStatus
$sel:status:LifecycleRule'

-> LifecycleRule

Create a value of LifecycleRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitions:LifecycleRule', lifecycleRule_transitions - Specifies when an Amazon S3 object transitions to a specified storage class.

$sel:noncurrentVersionExpiration:LifecycleRule', lifecycleRule_noncurrentVersionExpiration - Undocumented member.

$sel:prefix:LifecycleRule', lifecycleRule_prefix - Prefix identifying one or more objects to which the rule applies. This is no longer used; use Filter instead.

Replacement must be made for object keys containing special characters (such as carriage returns) when using XML requests. For more information, see XML related object key constraints.

$sel:noncurrentVersionTransitions:LifecycleRule', lifecycleRule_noncurrentVersionTransitions - Specifies the transition rule for the lifecycle rule that describes when noncurrent objects transition to a specific storage class. If your bucket is versioning-enabled (or versioning is suspended), you can set this action to request that Amazon S3 transition noncurrent object versions to a specific storage class at a set period in the object's lifetime.

$sel:expiration:LifecycleRule', lifecycleRule_expiration - Specifies the expiration for the lifecycle of the object in the form of date, days and, whether the object has a delete marker.

$sel:id:LifecycleRule', lifecycleRule_id - Unique identifier for the rule. The value cannot be longer than 255 characters.

$sel:filter':LifecycleRule', lifecycleRule_filter - The Filter is used to identify objects that a Lifecycle Rule applies to. A Filter must have exactly one of Prefix, Tag, or And specified. Filter is required if the LifecycleRule does not containt a Prefix element.

$sel:abortIncompleteMultipartUpload:LifecycleRule', lifecycleRule_abortIncompleteMultipartUpload - Undocumented member.

$sel:status:LifecycleRule', lifecycleRule_status - If 'Enabled', the rule is currently being applied. If 'Disabled', the rule is not currently being applied.

lifecycleRule_transitions :: Lens' LifecycleRule (Maybe [Transition])

Specifies when an Amazon S3 object transitions to a specified storage class.

lifecycleRule_noncurrentVersionExpiration :: Lens' LifecycleRule (Maybe NoncurrentVersionExpiration)

Undocumented member.

lifecycleRule_prefix :: Lens' LifecycleRule (Maybe Text)

Prefix identifying one or more objects to which the rule applies. This is no longer used; use Filter instead.

Replacement must be made for object keys containing special characters (such as carriage returns) when using XML requests. For more information, see XML related object key constraints.

lifecycleRule_noncurrentVersionTransitions :: Lens' LifecycleRule (Maybe [NoncurrentVersionTransition])

Specifies the transition rule for the lifecycle rule that describes when noncurrent objects transition to a specific storage class. If your bucket is versioning-enabled (or versioning is suspended), you can set this action to request that Amazon S3 transition noncurrent object versions to a specific storage class at a set period in the object's lifetime.

lifecycleRule_expiration :: Lens' LifecycleRule (Maybe LifecycleExpiration)

Specifies the expiration for the lifecycle of the object in the form of date, days and, whether the object has a delete marker.

lifecycleRule_id :: Lens' LifecycleRule (Maybe Text)

Unique identifier for the rule. The value cannot be longer than 255 characters.

lifecycleRule_filter :: Lens' LifecycleRule (Maybe LifecycleRuleFilter)

The Filter is used to identify objects that a Lifecycle Rule applies to. A Filter must have exactly one of Prefix, Tag, or And specified. Filter is required if the LifecycleRule does not containt a Prefix element.

lifecycleRule_abortIncompleteMultipartUpload :: Lens' LifecycleRule (Maybe AbortIncompleteMultipartUpload)

Undocumented member.

lifecycleRule_status :: Lens' LifecycleRule ExpirationStatus

If 'Enabled', the rule is currently being applied. If 'Disabled', the rule is not currently being applied.