Network.AWS.S3.Types.Owner

Description

Documentation

data Owner

Container for the owner's display name and ID.

See: newOwner smart constructor.

Constructors

Owner'

Fields

id :: Maybe Text

Container for the ID of the owner.

displayName :: Maybe Text

Container for the display name of the owner.

newOwner :: Owner

Create a value of Owner with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:Owner', owner_id - Container for the ID of the owner.

$sel:displayName:Owner', owner_displayName - Container for the display name of the owner.

owner_id :: Lens' Owner (Maybe Text)

Container for the ID of the owner.

owner_displayName :: Lens' Owner (Maybe Text)

Container for the display name of the owner.