Network.AWS.S3.Types.Part

Description

Documentation

data Part

Container for elements related to a part.

See: newPart smart constructor.

Constructors

Part'

Fields

eTag :: Maybe ETag

Entity tag returned when the part was uploaded.

size :: Maybe Integer

Size in bytes of the uploaded part data.

partNumber :: Maybe Int

Part number identifying the part. This is a positive integer between 1 and 10,000.

lastModified :: Maybe ISO8601

Date and time at which the part was uploaded.

newPart :: Part

Create a value of Part with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:Part', part_eTag - Entity tag returned when the part was uploaded.

$sel:size:Part', part_size - Size in bytes of the uploaded part data.

$sel:partNumber:Part', part_partNumber - Part number identifying the part. This is a positive integer between 1 and 10,000.

$sel:lastModified:Part', part_lastModified - Date and time at which the part was uploaded.

part_eTag :: Lens' Part (Maybe ETag)

Entity tag returned when the part was uploaded.

part_size :: Lens' Part (Maybe Integer)

Size in bytes of the uploaded part data.

part_partNumber :: Lens' Part (Maybe Int)

Part number identifying the part. This is a positive integer between 1 and 10,000.

part_lastModified :: Lens' Part (Maybe UTCTime)

Date and time at which the part was uploaded.