Network.AWS.S3.Types.Progress

Description

Documentation

data Progress

This data type contains information about progress of an operation.

See: newProgress smart constructor.

Constructors

Progress'

Fields

bytesReturned :: Maybe Integer

The current number of bytes of records payload data returned.

bytesScanned :: Maybe Integer

The current number of object bytes scanned.

bytesProcessed :: Maybe Integer

The current number of uncompressed object bytes processed.

newProgress :: Progress

Create a value of Progress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bytesReturned:Progress', progress_bytesReturned - The current number of bytes of records payload data returned.

$sel:bytesScanned:Progress', progress_bytesScanned - The current number of object bytes scanned.

$sel:bytesProcessed:Progress', progress_bytesProcessed - The current number of uncompressed object bytes processed.

progress_bytesReturned :: Lens' Progress (Maybe Integer)

The current number of bytes of records payload data returned.

progress_bytesScanned :: Lens' Progress (Maybe Integer)

The current number of object bytes scanned.

progress_bytesProcessed :: Lens' Progress (Maybe Integer)

The current number of uncompressed object bytes processed.