Network.AWS.S3.Types.Redirect

Description

Documentation

data Redirect

Specifies how requests are redirected. In the event of an error, you can specify a different error code to return.

See: newRedirect smart constructor.

Constructors

Redirect'

Fields

hostName :: Maybe Text

The host name to use in the redirect request.

protocol :: Maybe Protocol

Protocol to use when redirecting requests. The default is the protocol that is used in the original request.

httpRedirectCode :: Maybe Text

The HTTP redirect code to use on the response. Not required if one of the siblings is present.

replaceKeyWith :: Maybe Text

The specific object key to use in the redirect request. For example, redirect request to error.html. Not required if one of the siblings is present. Can be present only if ReplaceKeyPrefixWith is not provided.

Replacement must be made for object keys containing special characters (such as carriage returns) when using XML requests. For more information, see XML related object key constraints.

replaceKeyPrefixWith :: Maybe Text

The object key prefix to use in the redirect request. For example, to redirect requests for all pages with prefix docs/ (objects in the docs/ folder) to documents/, you can set a condition block with KeyPrefixEquals set to docs/ and in the Redirect set ReplaceKeyPrefixWith to /documents. Not required if one of the siblings is present. Can be present only if ReplaceKeyWith is not provided.

Replacement must be made for object keys containing special characters (such as carriage returns) when using XML requests. For more information, see XML related object key constraints.

newRedirect :: Redirect

Create a value of Redirect with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostName:Redirect', redirect_hostName - The host name to use in the redirect request.

$sel:protocol:Redirect', redirect_protocol - Protocol to use when redirecting requests. The default is the protocol that is used in the original request.

$sel:httpRedirectCode:Redirect', redirect_httpRedirectCode - The HTTP redirect code to use on the response. Not required if one of the siblings is present.

$sel:replaceKeyWith:Redirect', redirect_replaceKeyWith - The specific object key to use in the redirect request. For example, redirect request to error.html. Not required if one of the siblings is present. Can be present only if ReplaceKeyPrefixWith is not provided.

Replacement must be made for object keys containing special characters (such as carriage returns) when using XML requests. For more information, see XML related object key constraints.

$sel:replaceKeyPrefixWith:Redirect', redirect_replaceKeyPrefixWith - The object key prefix to use in the redirect request. For example, to redirect requests for all pages with prefix docs/ (objects in the docs/ folder) to documents/, you can set a condition block with KeyPrefixEquals set to docs/ and in the Redirect set ReplaceKeyPrefixWith to /documents. Not required if one of the siblings is present. Can be present only if ReplaceKeyWith is not provided.

Replacement must be made for object keys containing special characters (such as carriage returns) when using XML requests. For more information, see XML related object key constraints.

redirect_hostName :: Lens' Redirect (Maybe Text)

The host name to use in the redirect request.

redirect_protocol :: Lens' Redirect (Maybe Protocol)

Protocol to use when redirecting requests. The default is the protocol that is used in the original request.

redirect_httpRedirectCode :: Lens' Redirect (Maybe Text)

The HTTP redirect code to use on the response. Not required if one of the siblings is present.

redirect_replaceKeyWith :: Lens' Redirect (Maybe Text)

The specific object key to use in the redirect request. For example, redirect request to error.html. Not required if one of the siblings is present. Can be present only if ReplaceKeyPrefixWith is not provided.

Replacement must be made for object keys containing special characters (such as carriage returns) when using XML requests. For more information, see XML related object key constraints.

redirect_replaceKeyPrefixWith :: Lens' Redirect (Maybe Text)

The object key prefix to use in the redirect request. For example, to redirect requests for all pages with prefix docs/ (objects in the docs/ folder) to documents/, you can set a condition block with KeyPrefixEquals set to docs/ and in the Redirect set ReplaceKeyPrefixWith to /documents. Not required if one of the siblings is present. Can be present only if ReplaceKeyWith is not provided.

Replacement must be made for object keys containing special characters (such as carriage returns) when using XML requests. For more information, see XML related object key constraints.