Network.AWS.S3.Types.ReplicaModifications

Description

Documentation

data ReplicaModifications

A filter that you can specify for selection for modifications on replicas. Amazon S3 doesn't replicate replica modifications by default. In the latest version of replication configuration (when Filter is specified), you can specify this element and set the status to Enabled to replicate modifications on replicas.

If you don't specify the Filter element, Amazon S3 assumes that the replication configuration is the earlier version, V1. In the earlier version, this element is not allowed.

See: newReplicaModifications smart constructor.

Constructors

ReplicaModifications'

Fields

status :: ReplicaModificationsStatus

Specifies whether Amazon S3 replicates modifications on replicas.

newReplicaModifications

Arguments

:: ReplicaModificationsStatus
$sel:status:ReplicaModifications'

-> ReplicaModifications

Create a value of ReplicaModifications with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ReplicaModifications', replicaModifications_status - Specifies whether Amazon S3 replicates modifications on replicas.

replicaModifications_status :: Lens' ReplicaModifications ReplicaModificationsStatus

Specifies whether Amazon S3 replicates modifications on replicas.