Network.AWS.S3.Types.ReplicationRule

Description

Documentation

data ReplicationRule

Specifies which Amazon S3 objects to replicate and where to store the replicas.

See: newReplicationRule smart constructor.

Constructors

ReplicationRule'

Fields

deleteMarkerReplication :: Maybe DeleteMarkerReplication

priority :: Maybe Int

The priority indicates which rule has precedence whenever two or more replication rules conflict. Amazon S3 will attempt to replicate objects according to all replication rules. However, if there are two or more rules with the same destination bucket, then objects will be replicated according to the rule with the highest priority. The higher the number, the higher the priority.

For more information, see Replication in the Amazon S3 User Guide.

prefix :: Maybe Text

An object key name prefix that identifies the object or objects to which the rule applies. The maximum prefix length is 1,024 characters. To include all objects in a bucket, specify an empty string.

Replacement must be made for object keys containing special characters (such as carriage returns) when using XML requests. For more information, see XML related object key constraints.

existingObjectReplication :: Maybe ExistingObjectReplication

id :: Maybe Text

A unique identifier for the rule. The maximum value is 255 characters.

filter' :: Maybe ReplicationRuleFilter

sourceSelectionCriteria :: Maybe SourceSelectionCriteria

A container that describes additional filters for identifying the source objects that you want to replicate. You can choose to enable or disable the replication of these objects. Currently, Amazon S3 supports only the filter that you can specify for objects created with server-side encryption using a customer managed key stored in Amazon Web Services Key Management Service (SSE-KMS).

status :: ReplicationRuleStatus

Specifies whether the rule is enabled.

destination :: Destination

A container for information about the replication destination and its configurations including enabling the S3 Replication Time Control (S3 RTC).

newReplicationRule

Arguments

:: ReplicationRuleStatus
$sel:status:ReplicationRule'

-> Destination
$sel:destination:ReplicationRule'

-> ReplicationRule

Create a value of ReplicationRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deleteMarkerReplication:ReplicationRule', replicationRule_deleteMarkerReplication - Undocumented member.

$sel:priority:ReplicationRule', replicationRule_priority - The priority indicates which rule has precedence whenever two or more replication rules conflict. Amazon S3 will attempt to replicate objects according to all replication rules. However, if there are two or more rules with the same destination bucket, then objects will be replicated according to the rule with the highest priority. The higher the number, the higher the priority.

For more information, see Replication in the Amazon S3 User Guide.

$sel:prefix:ReplicationRule', replicationRule_prefix - An object key name prefix that identifies the object or objects to which the rule applies. The maximum prefix length is 1,024 characters. To include all objects in a bucket, specify an empty string.

Replacement must be made for object keys containing special characters (such as carriage returns) when using XML requests. For more information, see XML related object key constraints.

$sel:existingObjectReplication:ReplicationRule', replicationRule_existingObjectReplication -

$sel:id:ReplicationRule', replicationRule_id - A unique identifier for the rule. The maximum value is 255 characters.

$sel:filter':ReplicationRule', replicationRule_filter - Undocumented member.

$sel:sourceSelectionCriteria:ReplicationRule', replicationRule_sourceSelectionCriteria - A container that describes additional filters for identifying the source objects that you want to replicate. You can choose to enable or disable the replication of these objects. Currently, Amazon S3 supports only the filter that you can specify for objects created with server-side encryption using a customer managed key stored in Amazon Web Services Key Management Service (SSE-KMS).

$sel:status:ReplicationRule', replicationRule_status - Specifies whether the rule is enabled.

$sel:destination:ReplicationRule', replicationRule_destination - A container for information about the replication destination and its configurations including enabling the S3 Replication Time Control (S3 RTC).

replicationRule_deleteMarkerReplication :: Lens' ReplicationRule (Maybe DeleteMarkerReplication)

Undocumented member.

replicationRule_priority :: Lens' ReplicationRule (Maybe Int)

The priority indicates which rule has precedence whenever two or more replication rules conflict. Amazon S3 will attempt to replicate objects according to all replication rules. However, if there are two or more rules with the same destination bucket, then objects will be replicated according to the rule with the highest priority. The higher the number, the higher the priority.

For more information, see Replication in the Amazon S3 User Guide.

replicationRule_prefix :: Lens' ReplicationRule (Maybe Text)

An object key name prefix that identifies the object or objects to which the rule applies. The maximum prefix length is 1,024 characters. To include all objects in a bucket, specify an empty string.

Replacement must be made for object keys containing special characters (such as carriage returns) when using XML requests. For more information, see XML related object key constraints.

replicationRule_existingObjectReplication :: Lens' ReplicationRule (Maybe ExistingObjectReplication)

replicationRule_id :: Lens' ReplicationRule (Maybe Text)

A unique identifier for the rule. The maximum value is 255 characters.

replicationRule_filter :: Lens' ReplicationRule (Maybe ReplicationRuleFilter)

Undocumented member.

replicationRule_sourceSelectionCriteria :: Lens' ReplicationRule (Maybe SourceSelectionCriteria)

A container that describes additional filters for identifying the source objects that you want to replicate. You can choose to enable or disable the replication of these objects. Currently, Amazon S3 supports only the filter that you can specify for objects created with server-side encryption using a customer managed key stored in Amazon Web Services Key Management Service (SSE-KMS).

replicationRule_status :: Lens' ReplicationRule ReplicationRuleStatus

Specifies whether the rule is enabled.

replicationRule_destination :: Lens' ReplicationRule Destination

A container for information about the replication destination and its configurations including enabling the S3 Replication Time Control (S3 RTC).