Network.AWS.S3.Types.ReplicationTime

Description

Documentation

data ReplicationTime

A container specifying S3 Replication Time Control (S3 RTC) related information, including whether S3 RTC is enabled and the time when all objects and operations on objects must be replicated. Must be specified together with a Metrics block.

See: newReplicationTime smart constructor.

Constructors

ReplicationTime'

Fields

status :: ReplicationTimeStatus

Specifies whether the replication time is enabled.

time :: ReplicationTimeValue

A container specifying the time by which replication should be complete for all objects and operations on objects.

newReplicationTime

Arguments

:: ReplicationTimeStatus
$sel:status:ReplicationTime'

-> ReplicationTimeValue
$sel:time:ReplicationTime'

-> ReplicationTime

Create a value of ReplicationTime with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ReplicationTime', replicationTime_status - Specifies whether the replication time is enabled.

$sel:time:ReplicationTime', replicationTime_time - A container specifying the time by which replication should be complete for all objects and operations on objects.

replicationTime_status :: Lens' ReplicationTime ReplicationTimeStatus

Specifies whether the replication time is enabled.

replicationTime_time :: Lens' ReplicationTime ReplicationTimeValue

A container specifying the time by which replication should be complete for all objects and operations on objects.