Network.AWS.S3.Types.S3Location

Description

Documentation

data S3Location

Describes an Amazon S3 location that will receive the results of the restore request.

See: newS3Location smart constructor.

Constructors

S3Location'

Fields

cannedACL :: Maybe ObjectCannedACL

The canned ACL to apply to the restore results.

accessControlList :: Maybe [Grant]

A list of grants that control access to the staged results.

userMetadata :: Maybe [MetadataEntry]

A list of metadata to store with the restore results in S3.

encryption :: Maybe Encryption

storageClass :: Maybe StorageClass

The class of storage used to store the restore results.

tagging :: Maybe Tagging

The tag-set that is applied to the restore results.

bucketName :: BucketName

The name of the bucket where the restore results will be placed.

prefix :: Text

The prefix that is prepended to the restore results for this request.

newS3Location

Arguments

:: BucketName
$sel:bucketName:S3Location'

-> Text
$sel:prefix:S3Location'

-> S3Location

Create a value of S3Location with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cannedACL:S3Location', s3Location_cannedACL - The canned ACL to apply to the restore results.

$sel:accessControlList:S3Location', s3Location_accessControlList - A list of grants that control access to the staged results.

$sel:userMetadata:S3Location', s3Location_userMetadata - A list of metadata to store with the restore results in S3.

$sel:encryption:S3Location', s3Location_encryption - Undocumented member.

$sel:storageClass:S3Location', s3Location_storageClass - The class of storage used to store the restore results.

$sel:tagging:S3Location', s3Location_tagging - The tag-set that is applied to the restore results.

$sel:bucketName:S3Location', s3Location_bucketName - The name of the bucket where the restore results will be placed.

$sel:prefix:S3Location', s3Location_prefix - The prefix that is prepended to the restore results for this request.

s3Location_cannedACL :: Lens' S3Location (Maybe ObjectCannedACL)

The canned ACL to apply to the restore results.

s3Location_accessControlList :: Lens' S3Location (Maybe [Grant])

A list of grants that control access to the staged results.

s3Location_userMetadata :: Lens' S3Location (Maybe [MetadataEntry])

A list of metadata to store with the restore results in S3.

s3Location_encryption :: Lens' S3Location (Maybe Encryption)

Undocumented member.

s3Location_storageClass :: Lens' S3Location (Maybe StorageClass)

The class of storage used to store the restore results.

s3Location_tagging :: Lens' S3Location (Maybe Tagging)

The tag-set that is applied to the restore results.

s3Location_bucketName :: Lens' S3Location BucketName

The name of the bucket where the restore results will be placed.

s3Location_prefix :: Lens' S3Location Text

The prefix that is prepended to the restore results for this request.