Network.AWS.S3.Types.ServerSideEncryptionByDefault

Description

Documentation

data ServerSideEncryptionByDefault

Describes the default server-side encryption to apply to new objects in the bucket. If a PUT Object request doesn't specify any server-side encryption, this default encryption will be applied. For more information, see PUT Bucket encryption in the Amazon S3 API Reference.

See: newServerSideEncryptionByDefault smart constructor.

Constructors

ServerSideEncryptionByDefault'

Fields

kmsMasterKeyID :: Maybe (Sensitive Text)

Amazon Web Services Key Management Service (KMS) customer Amazon Web Services KMS key ID to use for the default encryption. This parameter is allowed if and only if SSEAlgorithm is set to aws:kms.

You can specify the key ID or the Amazon Resource Name (ARN) of the KMS key. However, if you are using encryption with cross-account operations, you must use a fully qualified KMS key ARN. For more information, see Using encryption for cross-account operations.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Amazon S3 only supports symmetric KMS keys and not asymmetric KMS keys. For more information, see Using symmetric and asymmetric keys in the Amazon Web Services Key Management Service Developer Guide.

sSEAlgorithm :: ServerSideEncryption

Server-side encryption algorithm to use for the default encryption.

newServerSideEncryptionByDefault

Arguments

:: ServerSideEncryption
$sel:sSEAlgorithm:ServerSideEncryptionByDefault'

-> ServerSideEncryptionByDefault

Create a value of ServerSideEncryptionByDefault with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsMasterKeyID:ServerSideEncryptionByDefault', serverSideEncryptionByDefault_kmsMasterKeyID - Amazon Web Services Key Management Service (KMS) customer Amazon Web Services KMS key ID to use for the default encryption. This parameter is allowed if and only if SSEAlgorithm is set to aws:kms.

You can specify the key ID or the Amazon Resource Name (ARN) of the KMS key. However, if you are using encryption with cross-account operations, you must use a fully qualified KMS key ARN. For more information, see Using encryption for cross-account operations.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Amazon S3 only supports symmetric KMS keys and not asymmetric KMS keys. For more information, see Using symmetric and asymmetric keys in the Amazon Web Services Key Management Service Developer Guide.

$sel:sSEAlgorithm:ServerSideEncryptionByDefault', serverSideEncryptionByDefault_sSEAlgorithm - Server-side encryption algorithm to use for the default encryption.

serverSideEncryptionByDefault_kmsMasterKeyID :: Lens' ServerSideEncryptionByDefault (Maybe Text)

Amazon Web Services Key Management Service (KMS) customer Amazon Web Services KMS key ID to use for the default encryption. This parameter is allowed if and only if SSEAlgorithm is set to aws:kms.

You can specify the key ID or the Amazon Resource Name (ARN) of the KMS key. However, if you are using encryption with cross-account operations, you must use a fully qualified KMS key ARN. For more information, see Using encryption for cross-account operations.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Amazon S3 only supports symmetric KMS keys and not asymmetric KMS keys. For more information, see Using symmetric and asymmetric keys in the Amazon Web Services Key Management Service Developer Guide.

serverSideEncryptionByDefault_sSEAlgorithm :: Lens' ServerSideEncryptionByDefault ServerSideEncryption

Server-side encryption algorithm to use for the default encryption.