Network.AWS.S3.Types.ServerSideEncryptionRule

Description

Documentation

data ServerSideEncryptionRule

Specifies the default server-side encryption configuration.

See: newServerSideEncryptionRule smart constructor.

Constructors

ServerSideEncryptionRule'

Fields

bucketKeyEnabled :: Maybe Bool

Specifies whether Amazon S3 should use an S3 Bucket Key with server-side encryption using KMS (SSE-KMS) for new objects in the bucket. Existing objects are not affected. Setting the BucketKeyEnabled element to true causes Amazon S3 to use an S3 Bucket Key. By default, S3 Bucket Key is not enabled.

For more information, see Amazon S3 Bucket Keys in the Amazon S3 User Guide.

applyServerSideEncryptionByDefault :: Maybe ServerSideEncryptionByDefault

Specifies the default server-side encryption to apply to new objects in the bucket. If a PUT Object request doesn't specify any server-side encryption, this default encryption will be applied.

newServerSideEncryptionRule :: ServerSideEncryptionRule

Create a value of ServerSideEncryptionRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucketKeyEnabled:ServerSideEncryptionRule', serverSideEncryptionRule_bucketKeyEnabled - Specifies whether Amazon S3 should use an S3 Bucket Key with server-side encryption using KMS (SSE-KMS) for new objects in the bucket. Existing objects are not affected. Setting the BucketKeyEnabled element to true causes Amazon S3 to use an S3 Bucket Key. By default, S3 Bucket Key is not enabled.

For more information, see Amazon S3 Bucket Keys in the Amazon S3 User Guide.

$sel:applyServerSideEncryptionByDefault:ServerSideEncryptionRule', serverSideEncryptionRule_applyServerSideEncryptionByDefault - Specifies the default server-side encryption to apply to new objects in the bucket. If a PUT Object request doesn't specify any server-side encryption, this default encryption will be applied.

serverSideEncryptionRule_bucketKeyEnabled :: Lens' ServerSideEncryptionRule (Maybe Bool)

Specifies whether Amazon S3 should use an S3 Bucket Key with server-side encryption using KMS (SSE-KMS) for new objects in the bucket. Existing objects are not affected. Setting the BucketKeyEnabled element to true causes Amazon S3 to use an S3 Bucket Key. By default, S3 Bucket Key is not enabled.

For more information, see Amazon S3 Bucket Keys in the Amazon S3 User Guide.

serverSideEncryptionRule_applyServerSideEncryptionByDefault :: Lens' ServerSideEncryptionRule (Maybe ServerSideEncryptionByDefault)

Specifies the default server-side encryption to apply to new objects in the bucket. If a PUT Object request doesn't specify any server-side encryption, this default encryption will be applied.