Network.AWS.S3.Types.TopicConfiguration

Description

Documentation

data TopicConfiguration

A container for specifying the configuration for publication of messages to an Amazon Simple Notification Service (Amazon SNS) topic when Amazon S3 detects specified events.

See: newTopicConfiguration smart constructor.

Constructors

TopicConfiguration'

Fields

id :: Maybe Text

filter' :: Maybe NotificationConfigurationFilter

topicArn :: Text

The Amazon Resource Name (ARN) of the Amazon SNS topic to which Amazon S3 publishes a message when it detects events of the specified type.

events :: [Event]

The Amazon S3 bucket event about which to send notifications. For more information, see Supported Event Types in the Amazon S3 User Guide.

newTopicConfiguration

Arguments

:: Text
$sel:topicArn:TopicConfiguration'

-> TopicConfiguration

Create a value of TopicConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:TopicConfiguration', topicConfiguration_id - Undocumented member.

$sel:filter':TopicConfiguration', topicConfiguration_filter - Undocumented member.

$sel:topicArn:TopicConfiguration', topicConfiguration_topicArn - The Amazon Resource Name (ARN) of the Amazon SNS topic to which Amazon S3 publishes a message when it detects events of the specified type.

$sel:events:TopicConfiguration', topicConfiguration_events - The Amazon S3 bucket event about which to send notifications. For more information, see Supported Event Types in the Amazon S3 User Guide.

topicConfiguration_id :: Lens' TopicConfiguration (Maybe Text)

Undocumented member.

topicConfiguration_filter :: Lens' TopicConfiguration (Maybe NotificationConfigurationFilter)

Undocumented member.

topicConfiguration_topicArn :: Lens' TopicConfiguration Text

The Amazon Resource Name (ARN) of the Amazon SNS topic to which Amazon S3 publishes a message when it detects events of the specified type.

topicConfiguration_events :: Lens' TopicConfiguration [Event]

The Amazon S3 bucket event about which to send notifications. For more information, see Supported Event Types in the Amazon S3 User Guide.