Network.AWS.S3.Types.Transition

Description

Documentation

data Transition

Specifies when an object transitions to a specified storage class. For more information about Amazon S3 lifecycle configuration rules, see Transitioning Objects Using Amazon S3 Lifecycle in the Amazon S3 User Guide.

See: newTransition smart constructor.

Constructors

Transition'

Fields

days :: Maybe Int

Indicates the number of days after creation when objects are transitioned to the specified storage class. The value must be a positive integer.

storageClass :: Maybe TransitionStorageClass

The storage class to which you want the object to transition.

date :: Maybe ISO8601

Indicates when objects are transitioned to the specified storage class. The date value must be in ISO 8601 format. The time is always midnight UTC.

newTransition :: Transition

Create a value of Transition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:days:Transition', transition_days - Indicates the number of days after creation when objects are transitioned to the specified storage class. The value must be a positive integer.

$sel:storageClass:Transition', transition_storageClass - The storage class to which you want the object to transition.

$sel:date:Transition', transition_date - Indicates when objects are transitioned to the specified storage class. The date value must be in ISO 8601 format. The time is always midnight UTC.

transition_days :: Lens' Transition (Maybe Int)

Indicates the number of days after creation when objects are transitioned to the specified storage class. The value must be a positive integer.

transition_storageClass :: Lens' Transition (Maybe TransitionStorageClass)

The storage class to which you want the object to transition.

transition_date :: Lens' Transition (Maybe UTCTime)

Indicates when objects are transitioned to the specified storage class. The date value must be in ISO 8601 format. The time is always midnight UTC.