Network.AWS.S3.V2006_03_01.CopyObject

Description

Creates a copy of an object that is already stored in Amazon S3.

Documentation

type PutObjectCopy = CopyObject

copyObject

Arguments

:: Text
_corCopySource

-> BucketName
_corBucket

-> ObjectKey
_corKey

-> CopyObject

Minimum specification for a CopyObject request.

data CopyObject

Constructors

CopyObject

Fields

_corCopySource :: Text

The name of the source bucket and key name of the source object, separated by a slash (/). Must be URL-encoded.

_corBucket :: BucketName

_corKey :: ObjectKey

_corMetadata :: HashMap Text Text

A map of metadata to store with the object in S3.

_corCacheControl :: Maybe Text

Specifies caching behavior along the request/reply chain.

_corContentDisposition :: Maybe Text

Specifies presentational information for the object.

_corContentEncoding :: Maybe Text

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

_corContentLanguage :: Maybe Text

The language the content is in.

_corContentType :: Maybe Text

A standard MIME type describing the format of the object data.

_corCopySourceIfMatch :: Maybe Text

Copies the object if its entity tag (ETag) matches the specified tag.

_corCopySourceIfModifiedSince :: Maybe RFC822

Copies the object if it has been modified since the specified time.

_corCopySourceIfNoneMatch :: Maybe Text

Copies the object if its entity tag (ETag) is different than the specified ETag.

_corCopySourceIfUnmodifiedSince :: Maybe RFC822

Copies the object if it hasn't been modified since the specified time.

_corCopySourceSSECustomerAlgorithm :: Maybe Text

Specifies the algorithm to use when decrypting the source object (e.g., AES256).

_corCopySourceSSECustomerKey :: Maybe Text

Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

_corCopySourceSSECustomerKeyMD5 :: Maybe Text

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

_corExpires :: Maybe RFC822

The date and time at which the object is no longer cacheable.

_corGrantFullControl :: Maybe Text

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

_corGrantRead :: Maybe Text

Allows grantee to read the object data and its metadata.

_corGrantReadACP :: Maybe Text

Allows grantee to read the object ACL.

_corGrantWriteACP :: Maybe Text

Allows grantee to write the ACL for the applicable object.

_corMetadataDirective :: Maybe MetadataDirective

Specifies whether the metadata is copied from the source object or replaced with metadata provided in the request.

_corACL :: Maybe ObjectCannedACL

The canned ACL to apply to the object.

_corSSECustomerAlgorithm :: Maybe Text

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

_corSSECustomerKey :: Maybe Text

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

_corSSECustomerKeyMD5 :: Maybe Text

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

_corServerSideEncryption :: Maybe ServerSideEncryption

The Server-side encryption algorithm used when storing this object in S3.

_corStorageClass :: Maybe StorageClass

The type of storage to use for the object. Defaults to STANDARD.

_corWebsiteRedirectLocation :: Maybe Text

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

data CopyObjectResponse

Constructors

CopyObjectResponse

Fields

_cooCopyObjectResult :: Maybe CopyObjectResult

_cooCopySourceVersionId :: Maybe Text

_cooExpiration :: Maybe RFC822

If the object expiration is configured, the response includes this header.

_cooSSECustomerAlgorithm :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

_cooSSECustomerKeyMD5 :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

_cooServerSideEncryption :: Maybe ServerSideEncryption

The Server-side encryption algorithm used when storing this object in S3.