Network.AWS.S3.V2006_03_01.CreateBucket

Description

Creates a new bucket.

Documentation

type PutBucket = CreateBucket

createBucket

Arguments

:: BucketName
_cbrBucket

-> CreateBucketConfiguration
_cbrCreateBucketConfiguration

-> CreateBucket

Minimum specification for a CreateBucket request.

data CreateBucket

Constructors

CreateBucket

Fields

_cbrBucket :: BucketName

_cbrCreateBucketConfiguration :: CreateBucketConfiguration

_cbrACL :: Maybe BucketCannedACL

The canned ACL to apply to the bucket.

_cbrGrantFullControl :: Maybe Text

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

_cbrGrantRead :: Maybe Text

Allows grantee to list the objects in the bucket.

_cbrGrantReadACP :: Maybe Text

Allows grantee to read the bucket ACL.

_cbrGrantWrite :: Maybe Text

Allows grantee to create, overwrite, and delete any object in the bucket.

_cbrGrantWriteACP :: Maybe Text

Allows grantee to write the ACL for the applicable bucket.

data CreateBucketResponse

Constructors

CreateBucketResponse

Fields

_cboLocation :: Maybe Text