Network.AWS.S3.V2006_03_01.DeleteObject

Description

Removes the null version (if there is one) of an object and inserts a delete marker, which becomes the latest version of the object. If there isn't a null version, Amazon S3 does not remove any objects.

Documentation

deleteObject

Arguments

:: BucketName
_dorBucket

-> ObjectKey
_dorKey

-> DeleteObject

Minimum specification for a DeleteObject request.

data DeleteObject

Constructors

DeleteObject

Fields

_dorBucket :: BucketName

_dorKey :: ObjectKey

_dorMFA :: Maybe Text

The concatenation of the authentication device's serial number, a space, and the value that is displayed on your authentication device.

_dorVersionId :: Maybe ObjectVersionId

VersionId used to reference a specific version of the object.

data DeleteObjectResponse

Constructors

DeleteObjectResponse

Fields

_dooDeleteMarker :: Maybe Bool

Specifies whether the versioned object that was permanently deleted was (true) or was not (false) a delete marker.

_dooVersionId :: Maybe ObjectVersionId

Returns the version ID of the delete marker created as a result of the DELETE operation.