Network.AWS.S3.V2006_03_01.GetBucketVersioning

Description

Returns the versioning state of a bucket.

Documentation

getBucketVersioning

Arguments

:: BucketName
_gbvrBucket

-> GetBucketVersioning

Default GetBucketVersioning request.

data GetBucketVersioning

Constructors

GetBucketVersioning

Fields

_gbvrBucket :: BucketName

data GetBucketVersioningResponse

Constructors

GetBucketVersioningResponse

Fields

_gbvoStatus :: Maybe (Switch BucketVersioningStatus)

The versioning state of the bucket.

_gbvoMfaDelete :: Maybe (Switch MFADeleteStatus)

Specifies whether MFA delete is enabled in the bucket versioning configuration. This element is only returned if the bucket has been configured with MFA delete. If the bucket has never been so configured, this element is not returned.