Network.AWS.S3.V2006_03_01.ListMultipartUploads

Description

This operation lists in-progress multipart uploads.

Documentation

listMultipartUploads

Arguments

:: BucketName
_lmurBucket

-> ListMultipartUploads

Minimum specification for a ListMultipartUploads request.

data ListMultipartUploads

Constructors

ListMultipartUploads

Fields

_lmurBucket :: BucketName

_lmurDelimiter :: Maybe Char

Character you use to group keys.

_lmurEncodingType :: Maybe EncodingType

Requests Amazon S3 to encode the object keys in the response and specifies the encoding method to use. An object key may contain any Unicode character; however, XML 1.0 parser cannot parse some characters, such as characters with an ASCII value from 0 to 10. For characters that are not supported in XML 1.0, you can add this parameter to request that Amazon S3 encode the keys in the response.

_lmurKeyMarker :: Maybe Text

Together with upload-id-marker, this parameter specifies the multipart upload after which listing should begin.

_lmurMaxUploads :: Maybe Integer

Sets the maximum number of multipart uploads, from 1 to 1,000, to return in the response body. 1,000 is the maximum number of uploads that can be returned in a response.

_lmurPrefix :: Maybe Text

Lists in-progress uploads only for those keys that begin with the specified prefix.

_lmurUploadIdMarker :: Maybe Text

Together with key-marker, specifies the multipart upload after which listing should begin. If key-marker is not specified, the upload-id-marker parameter is ignored.

data ListMultipartUploadsResponse

Constructors

ListMultipartUploadsResponse

Fields

_lmuoIsTruncated :: Bool

Indicates whether the returned list of multipart uploads is truncated. A value of true indicates that the list was truncated. The list can be truncated if the number of multipart uploads exceeds the limit allowed or specified by max uploads.

_lmuoBucket :: Maybe BucketName

Name of the bucket to which the multipart upload was initiated.

_lmuoCommonPrefixes :: [CommonPrefix]

_lmuoEncodingType :: Maybe EncodingType

Encoding type used by Amazon S3 to encode object keys in the response.

_lmuoKeyMarker :: Maybe Text

The key at or after which the listing began.

_lmuoMaxUploads :: Maybe Integer

Maximum number of multipart uploads that could have been included in the response.

_lmuoUploads :: [MultipartUpload]

_lmuoNextKeyMarker :: Maybe Text

When a list is truncated, this element specifies the value that should be used for the key-marker request parameter in a subsequent request.

_lmuoNextUploadIdMarker :: Maybe Text

When a list is truncated, this element specifies the value that should be used for the upload-id-marker request parameter in a subsequent request.

_lmuoPrefix :: Maybe Text

When a prefix is provided in the request, this field contains the specified prefix. The result contains only keys starting with the specified prefix.

_lmuoUploadIdMarker :: Maybe Text

Upload ID after which listing began.