Network.AWS.S3.V2006_03_01.ListParts

Description

Lists the parts that have been uploaded for a specific multipart upload.

Documentation

listParts

Arguments

:: BucketName
_lprBucket

-> Text
_lprUploadId

-> ObjectKey
_lprKey

-> ListParts

Minimum specification for a ListParts request.

data ListParts

Constructors

ListParts

Fields

_lprBucket :: BucketName

_lprUploadId :: Text

Upload ID identifying the multipart upload whose parts are being listed.

_lprKey :: ObjectKey

_lprMaxParts :: Maybe Integer

Sets the maximum number of parts to return.

_lprPartNumberMarker :: Maybe Integer

Specifies the part after which listing should begin. Only parts with higher part numbers will be listed.

lprUploadId :: Lens' ListParts Text

lprPartNumberMarker :: Lens' ListParts (Maybe Integer)

lprMaxParts :: Lens' ListParts (Maybe Integer)

lprKey :: Lens' ListParts ObjectKey

lprBucket :: Lens' ListParts BucketName

data ListPartsResponse

Constructors

ListPartsResponse

Fields

_lpoIsTruncated :: Bool

Indicates whether the returned list of parts is truncated.

_lpoBucket :: Maybe BucketName

Name of the bucket to which the multipart upload was initiated.

_lpoInitiator :: Maybe Initiator

Identifies who initiated the multipart upload.

_lpoMaxParts :: Maybe Integer

Maximum number of parts that were allowed in the response.

_lpoUploadId :: Maybe Text

Upload ID identifying the multipart upload whose parts are being listed.

_lpoNextPartNumberMarker :: Maybe Integer

When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

_lpoKey :: Maybe ObjectKey

Object key for which the multipart upload was initiated.

_lpoOwner :: Maybe Owner

_lpoPartNumberMarker :: Maybe Integer

Part number after which listing begins.

_lpoParts :: [Part]

_lpoStorageClass :: Maybe StorageClass

The class of storage used to store the object.

lpoUploadId :: Lens' ListPartsResponse (Maybe Text)

lpoStorageClass :: Lens' ListPartsResponse (Maybe StorageClass)

lpoParts :: Lens' ListPartsResponse [Part]

lpoPartNumberMarker :: Lens' ListPartsResponse (Maybe Integer)

lpoOwner :: Lens' ListPartsResponse (Maybe Owner)

lpoNextPartNumberMarker :: Lens' ListPartsResponse (Maybe Integer)

lpoMaxParts :: Lens' ListPartsResponse (Maybe Integer)

lpoKey :: Lens' ListPartsResponse (Maybe ObjectKey)

lpoIsTruncated :: Lens' ListPartsResponse Bool

lpoInitiator :: Lens' ListPartsResponse (Maybe Initiator)

lpoBucket :: Lens' ListPartsResponse (Maybe BucketName)