Network.AWS.S3.V2006_03_01.PutBucketLogging

Description

Set the logging parameters for a bucket and to specify permissions for who can view and modify the logging parameters. To set the logging status of a bucket, you must be the bucket owner.

Documentation

putBucketLogging

Arguments

:: BucketLoggingStatus
_pblrBucketLoggingStatus

-> BucketName
_pblrBucket

-> PutBucketLogging

Minimum specification for a PutBucketLogging request.

data PutBucketLogging

Constructors

PutBucketLogging

Fields

_pblrBucketLoggingStatus :: BucketLoggingStatus

_pblrBucket :: BucketName

_pblrContentMD5 :: Maybe Text

pblrContentMD5 :: Lens' PutBucketLogging (Maybe Text)

pblrBucketLoggingStatus :: Lens' PutBucketLogging BucketLoggingStatus

pblrBucket :: Lens' PutBucketLogging BucketName

data PutBucketLoggingResponse

Constructors

PutBucketLoggingResponse