Network.AWS.S3.V2006_03_01.PutBucketRequestPayment

Description

Sets the request payment configuration for a bucket. By default, the bucket owner pays for downloads from the bucket. This configuration parameter enables the bucket owner (only) to specify that the person requesting the download will be charged for the download.

Documentation

putBucketRequestPayment

Arguments

:: RequestPaymentConfiguration
_pbrprRequestPaymentConfiguration

-> BucketName
_pbrprBucket

-> PutBucketRequestPayment

Minimum specification for a PutBucketRequestPayment request.

data PutBucketRequestPayment

Constructors

PutBucketRequestPayment

Fields

_pbrprRequestPaymentConfiguration :: RequestPaymentConfiguration

_pbrprBucket :: BucketName

_pbrprContentMD5 :: Maybe Text

data PutBucketRequestPaymentResponse

Constructors

PutBucketRequestPaymentResponse