Network.AWS.S3.V2006_03_01.PutObject

Description

Adds an object to a bucket.

Documentation

putObject

Arguments

:: BucketName
_porBucket

-> ObjectKey
_porKey

-> RqBody
_porBody

-> PutObject

Minimum specification for a PutObject request.

data PutObject

Constructors

PutObject

Fields

_porBucket :: BucketName

_porKey :: ObjectKey

_porBody :: RqBody

_porMetadata :: HashMap Text Text

A map of metadata to store with the object in S3.

_porCacheControl :: Maybe Text

Specifies caching behavior along the request/reply chain.

_porContentDisposition :: Maybe Text

Specifies presentational information for the object.

_porContentEncoding :: Maybe Text

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

_porContentLanguage :: Maybe Text

The language the content is in.

_porContentLength :: Maybe Integer

Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically.

_porContentMD5 :: Maybe Text

_porContentType :: Maybe Text

A standard MIME type describing the format of the object data.

_porExpires :: Maybe RFC822

The date and time at which the object is no longer cacheable.

_porGrantFullControl :: Maybe Text

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

_porGrantRead :: Maybe Text

Allows grantee to read the object data and its metadata.

_porGrantReadACP :: Maybe Text

Allows grantee to read the object ACL.

_porGrantWriteACP :: Maybe Text

Allows grantee to write the ACL for the applicable object.

_porACL :: Maybe ObjectCannedACL

The canned ACL to apply to the object.

_porSSECustomerAlgorithm :: Maybe Text

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

_porSSECustomerKey :: Maybe Text

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

_porSSECustomerKeyMD5 :: Maybe Text

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

_porServerSideEncryption :: Maybe ServerSideEncryption

The Server-side encryption algorithm used when storing this object in S3.

_porStorageClass :: Maybe StorageClass

The type of storage to use for the object. Defaults to STANDARD.

_porWebsiteRedirectLocation :: Maybe Text

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

data PutObjectResponse

Constructors

PutObjectResponse

Fields

_pooETag :: Maybe ETag

Entity tag for the uploaded object.

_pooExpiration :: Maybe RFC822

If the object expiration is configured, this will contain the expiration date (expiry-date) and rule ID (rule-id). The value of rule-id is URL encoded.

_pooVersionId :: Maybe ObjectVersionId

Version of the object.

_pooSSECustomerAlgorithm :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

_pooSSECustomerKeyMD5 :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

_pooServerSideEncryption :: Maybe ServerSideEncryption

The Server-side encryption algorithm used when storing this object in S3.