Network.AWS.S3.V2006_03_01.PutObjectAcl

Description

uses the acl subresource to set the access control list (ACL) permissions for an object that already exists in a bucket.

Documentation

putObjectAcl

Arguments

:: BucketName
_poarBucket

-> ObjectKey
_poarKey

-> PutObjectAcl

Minimum specification for a PutObjectAcl request.

data PutObjectAcl

Constructors

PutObjectAcl

Fields

_poarBucket :: BucketName

_poarKey :: ObjectKey

_poarAccessControlPolicy :: Maybe AccessControlPolicy

_poarContentMD5 :: Maybe Text

_poarGrantFullControl :: Maybe Text

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

_poarGrantRead :: Maybe Text

Allows grantee to list the objects in the bucket.

_poarGrantReadACP :: Maybe Text

Allows grantee to read the bucket ACL.

_poarGrantWrite :: Maybe Text

Allows grantee to create, overwrite, and delete any object in the bucket.

_poarGrantWriteACP :: Maybe Text

Allows grantee to write the ACL for the applicable bucket.

_poarACL :: Maybe ObjectCannedACL

The canned ACL to apply to the object.

poarKey :: Lens' PutObjectAcl ObjectKey

poarGrantWriteACP :: Lens' PutObjectAcl (Maybe Text)

poarGrantWrite :: Lens' PutObjectAcl (Maybe Text)

poarGrantReadACP :: Lens' PutObjectAcl (Maybe Text)

poarGrantRead :: Lens' PutObjectAcl (Maybe Text)

poarGrantFullControl :: Lens' PutObjectAcl (Maybe Text)

poarContentMD5 :: Lens' PutObjectAcl (Maybe Text)

poarBucket :: Lens' PutObjectAcl BucketName

poarAccessControlPolicy :: Lens' PutObjectAcl (Maybe AccessControlPolicy)

poarACL :: Lens' PutObjectAcl (Maybe ObjectCannedACL)

data PutObjectAclResponse

Constructors

PutObjectAclResponse