Network.AWS.S3.V2006_03_01.UploadPart

Description

Uploads a part in a multipart upload. Note: After you initiate multipart upload and upload one or more parts, you must either complete or abort multipart upload in order to stop getting charged for storage of the uploaded parts. Only after you either complete or abort multipart upload, Amazon S3 frees up the parts storage and stops charging you for the parts storage.

Documentation

uploadPart

Arguments

:: BucketName
_uprBucket

-> Text
_uprUploadId

-> ObjectKey
_uprKey

-> Integer
_uprPartNumber

-> RqBody
_uprBody

-> UploadPart

Default UploadPart request.

data UploadPart

Constructors

UploadPart

Fields

_uprBucket :: BucketName

_uprUploadId :: Text

Upload ID identifying the multipart upload whose part is being uploaded.

_uprKey :: ObjectKey

_uprPartNumber :: Integer

Part number of part being uploaded.

_uprBody :: RqBody

_uprContentLength :: Maybe Integer

Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically.

_uprContentMD5 :: Maybe Text

_uprSSECustomerAlgorithm :: Maybe Text

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

_uprSSECustomerKey :: Maybe Text

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

_uprSSECustomerKeyMD5 :: Maybe Text

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

data UploadPartResponse

Constructors

UploadPartResponse

Fields

_upoETag :: Maybe ETag

Entity tag for the uploaded object.

_upoSSECustomerAlgorithm :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

_upoSSECustomerKeyMD5 :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

_upoServerSideEncryption :: Maybe ServerSideEncryption

The Server-side encryption algorithm used when storing this object in S3.