Network.AWS.S3.WriteGetObjectResponse

Description

Passes transformed objects to a GetObject operation when using Object Lambda access points. For information about Object Lambda access points, see Transforming objects with Object Lambda access points in the Amazon S3 User Guide.

This operation supports metadata that can be returned by GetObject, in addition to RequestRoute, RequestToken, StatusCode, ErrorCode, and ErrorMessage. The GetObject response metadata is supported so that the WriteGetObjectResponse caller, typically an Lambda function, can provide the same metadata when it internally invokes GetObject. When WriteGetObjectResponse is called by a customer-owned Lambda function, the metadata returned to the end user GetObject call might differ from what Amazon S3 would normally return.

You can include any number of metadata headers. When including a metadata header, it should be prefaced with x-amz-meta. For example, x-amz-meta-my-custom-header: MyCustomValue. The primary use case for this is to forward GetObject metadata.

Amazon Web Services provides some prebuilt Lambda functions that you can use with S3 Object Lambda to detect and redact personally identifiable information (PII) and decompress S3 objects. These Lambda functions are available in the Amazon Web Services Serverless Application Repository, and can be selected through the Amazon Web Services Management Console when you create your Object Lambda access point.

Example 1: PII Access Control - This Lambda function uses Amazon Comprehend, a natural language processing (NLP) service using machine learning to find insights and relationships in text. It automatically detects personally identifiable information (PII) such as names, addresses, dates, credit card numbers, and social security numbers from documents in your Amazon S3 bucket.

Example 2: PII Redaction - This Lambda function uses Amazon Comprehend, a natural language processing (NLP) service using machine learning to find insights and relationships in text. It automatically redacts personally identifiable information (PII) such as names, addresses, dates, credit card numbers, and social security numbers from documents in your Amazon S3 bucket.

Example 3: Decompression - The Lambda function S3ObjectLambdaDecompression, is equipped to decompress objects stored in S3 in one of six compressed file formats including bzip2, gzip, snappy, zlib, zstandard and ZIP.

For information on how to view and use these functions, see Using Amazon Web Services built Lambda functions in the Amazon S3 User Guide.

Creating a Request

data WriteGetObjectResponse

See: newWriteGetObjectResponse smart constructor.

Constructors

WriteGetObjectResponse'

Fields

partsCount :: Maybe Int

The count of parts this object has.

requestCharged :: Maybe RequestCharged

eTag :: Maybe ETag

An opaque identifier assigned by a web server to a specific version of a resource found at a URL.

tagCount :: Maybe Int

The number of tags, if any, on the object.

contentType :: Maybe Text

A standard MIME type describing the format of the object data.

contentDisposition :: Maybe Text

Specifies presentational information for the object.

contentRange :: Maybe Text

The portion of the object returned in the response.

deleteMarker :: Maybe Bool

Specifies whether an object stored in Amazon S3 is (true) or is not (false) a delete marker.

contentLanguage :: Maybe Text

The language the content is in.

expiration :: Maybe Text

If object stored in Amazon S3 expiration is configured (see PUT Bucket lifecycle) it includes expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL encoded.

replicationStatus :: Maybe ReplicationStatus

Indicates if request involves bucket that is either a source or destination in a Replication rule. For more information about S3 Replication, see Replication.

contentLength :: Maybe Integer

The size of the content body in bytes.

metadata :: HashMap Text Text

A map of metadata to store with the object in S3.

sSECustomerKeyMD5 :: Maybe Text

128-bit MD5 digest of customer-provided encryption key used in Amazon S3 to encrypt data stored in S3. For more information, see Protecting data using server-side encryption with customer-provided encryption keys (SSE-C).

sSEKMSKeyId :: Maybe (Sensitive Text)

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for stored in Amazon S3 object.

contentEncoding :: Maybe Text

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

versionId :: Maybe ObjectVersionId

An ID used to reference a specific version of the object.

storageClass :: Maybe StorageClass

The class of storage used to store object in Amazon S3.

bucketKeyEnabled :: Maybe Bool

Indicates whether the object stored in Amazon S3 uses an S3 bucket key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

acceptRanges :: Maybe Text

Indicates that a range of bytes was specified.

missingMeta :: Maybe Int

Set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

statusCode :: Maybe Int

The integer status code for an HTTP response of a corresponding GetObject request.

Status Codes

200 - OK

206 - Partial Content

304 - Not Modified

400 - Bad Request

401 - Unauthorized

403 - Forbidden

404 - Not Found

405 - Method Not Allowed

409 - Conflict

411 - Length Required

412 - Precondition Failed

416 - Range Not Satisfiable

500 - Internal Server Error

503 - Service Unavailable

serverSideEncryption :: Maybe ServerSideEncryption

The server-side encryption algorithm used when storing requested object in Amazon S3 (for example, AES256, aws:kms).

sSECustomerAlgorithm :: Maybe Text

Encryption algorithm used if server-side encryption with a customer-provided encryption key was specified for object stored in Amazon S3.

objectLockLegalHoldStatus :: Maybe ObjectLockLegalHoldStatus

Indicates whether an object stored in Amazon S3 has an active legal hold.

cacheControl :: Maybe Text

Specifies caching behavior along the request/reply chain.

errorMessage :: Maybe Text

Contains a generic description of the error condition. Returned in the <Message> tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in body.

lastModified :: Maybe ISO8601

The date and time that the object was last modified.

expires :: Maybe ISO8601

The date and time at which the object is no longer cacheable.

restore :: Maybe Text

Provides information about object restoration operation and expiration time of the restored object copy.

objectLockMode :: Maybe ObjectLockMode

Indicates whether an object stored in Amazon S3 has Object Lock enabled. For more information about S3 Object Lock, see Object Lock.

errorCode :: Maybe Text

A string that uniquely identifies an error condition. Returned in the <Code> tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in the body. All error codes from S3 are sentence-cased. Regex value is "^[A-Z][a-zA-Z]+$".

objectLockRetainUntilDate :: Maybe ISO8601

The date and time when Object Lock is configured to expire.

requestRoute :: Text

Route prefix to the HTTP URL generated.

requestToken :: Text

A single use encrypted token that maps WriteGetObjectResponse to the end user GetObject request.

body :: RequestBody

The object data.

newWriteGetObjectResponse

Arguments

:: Text
$sel:requestRoute:WriteGetObjectResponse'

-> Text
$sel:requestToken:WriteGetObjectResponse'

-> RequestBody
$sel:body:WriteGetObjectResponse'

-> WriteGetObjectResponse

Create a value of WriteGetObjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:partsCount:WriteGetObjectResponse', writeGetObjectResponse_partsCount - The count of parts this object has.

$sel:requestCharged:WriteGetObjectResponse', writeGetObjectResponse_requestCharged - Undocumented member.

$sel:eTag:WriteGetObjectResponse', writeGetObjectResponse_eTag - An opaque identifier assigned by a web server to a specific version of a resource found at a URL.

$sel:tagCount:WriteGetObjectResponse', writeGetObjectResponse_tagCount - The number of tags, if any, on the object.

$sel:contentType:WriteGetObjectResponse', writeGetObjectResponse_contentType - A standard MIME type describing the format of the object data.

$sel:contentDisposition:WriteGetObjectResponse', writeGetObjectResponse_contentDisposition - Specifies presentational information for the object.

$sel:contentRange:WriteGetObjectResponse', writeGetObjectResponse_contentRange - The portion of the object returned in the response.

$sel:deleteMarker:WriteGetObjectResponse', writeGetObjectResponse_deleteMarker - Specifies whether an object stored in Amazon S3 is (true) or is not (false) a delete marker.

$sel:contentLanguage:WriteGetObjectResponse', writeGetObjectResponse_contentLanguage - The language the content is in.

$sel:expiration:WriteGetObjectResponse', writeGetObjectResponse_expiration - If object stored in Amazon S3 expiration is configured (see PUT Bucket lifecycle) it includes expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL encoded.

$sel:replicationStatus:WriteGetObjectResponse', writeGetObjectResponse_replicationStatus - Indicates if request involves bucket that is either a source or destination in a Replication rule. For more information about S3 Replication, see Replication.

$sel:contentLength:WriteGetObjectResponse', writeGetObjectResponse_contentLength - The size of the content body in bytes.

$sel:metadata:WriteGetObjectResponse', writeGetObjectResponse_metadata - A map of metadata to store with the object in S3.

$sel:sSECustomerKeyMD5:WriteGetObjectResponse', writeGetObjectResponse_sSECustomerKeyMD5 - 128-bit MD5 digest of customer-provided encryption key used in Amazon S3 to encrypt data stored in S3. For more information, see Protecting data using server-side encryption with customer-provided encryption keys (SSE-C).

$sel:sSEKMSKeyId:WriteGetObjectResponse', writeGetObjectResponse_sSEKMSKeyId - If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for stored in Amazon S3 object.

$sel:contentEncoding:WriteGetObjectResponse', writeGetObjectResponse_contentEncoding - Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

$sel:versionId:WriteGetObjectResponse', writeGetObjectResponse_versionId - An ID used to reference a specific version of the object.

$sel:storageClass:WriteGetObjectResponse', writeGetObjectResponse_storageClass - The class of storage used to store object in Amazon S3.

$sel:bucketKeyEnabled:WriteGetObjectResponse', writeGetObjectResponse_bucketKeyEnabled - Indicates whether the object stored in Amazon S3 uses an S3 bucket key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

$sel:acceptRanges:WriteGetObjectResponse', writeGetObjectResponse_acceptRanges - Indicates that a range of bytes was specified.

$sel:missingMeta:WriteGetObjectResponse', writeGetObjectResponse_missingMeta - Set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

$sel:statusCode:WriteGetObjectResponse', writeGetObjectResponse_statusCode - The integer status code for an HTTP response of a corresponding GetObject request.

Status Codes

200 - OK

206 - Partial Content

304 - Not Modified

400 - Bad Request

401 - Unauthorized

403 - Forbidden

404 - Not Found

405 - Method Not Allowed

409 - Conflict

411 - Length Required

412 - Precondition Failed

416 - Range Not Satisfiable

500 - Internal Server Error

503 - Service Unavailable

$sel:serverSideEncryption:WriteGetObjectResponse', writeGetObjectResponse_serverSideEncryption - The server-side encryption algorithm used when storing requested object in Amazon S3 (for example, AES256, aws:kms).

$sel:sSECustomerAlgorithm:WriteGetObjectResponse', writeGetObjectResponse_sSECustomerAlgorithm - Encryption algorithm used if server-side encryption with a customer-provided encryption key was specified for object stored in Amazon S3.

$sel:objectLockLegalHoldStatus:WriteGetObjectResponse', writeGetObjectResponse_objectLockLegalHoldStatus - Indicates whether an object stored in Amazon S3 has an active legal hold.

$sel:cacheControl:WriteGetObjectResponse', writeGetObjectResponse_cacheControl - Specifies caching behavior along the request/reply chain.

$sel:errorMessage:WriteGetObjectResponse', writeGetObjectResponse_errorMessage - Contains a generic description of the error condition. Returned in the <Message> tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in body.

$sel:lastModified:WriteGetObjectResponse', writeGetObjectResponse_lastModified - The date and time that the object was last modified.

$sel:expires:WriteGetObjectResponse', writeGetObjectResponse_expires - The date and time at which the object is no longer cacheable.

$sel:restore:WriteGetObjectResponse', writeGetObjectResponse_restore - Provides information about object restoration operation and expiration time of the restored object copy.

$sel:objectLockMode:WriteGetObjectResponse', writeGetObjectResponse_objectLockMode - Indicates whether an object stored in Amazon S3 has Object Lock enabled. For more information about S3 Object Lock, see Object Lock.

$sel:errorCode:WriteGetObjectResponse', writeGetObjectResponse_errorCode - A string that uniquely identifies an error condition. Returned in the <Code> tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in the body. All error codes from S3 are sentence-cased. Regex value is "^[A-Z][a-zA-Z]+$".

$sel:objectLockRetainUntilDate:WriteGetObjectResponse', writeGetObjectResponse_objectLockRetainUntilDate - The date and time when Object Lock is configured to expire.

$sel:requestRoute:WriteGetObjectResponse', writeGetObjectResponse_requestRoute - Route prefix to the HTTP URL generated.

$sel:requestToken:WriteGetObjectResponse', writeGetObjectResponse_requestToken - A single use encrypted token that maps WriteGetObjectResponse to the end user GetObject request.

$sel:body:WriteGetObjectResponse', writeGetObjectResponse_body - The object data.

Request Lenses

writeGetObjectResponse_partsCount :: Lens' WriteGetObjectResponse (Maybe Int)

The count of parts this object has.

writeGetObjectResponse_requestCharged :: Lens' WriteGetObjectResponse (Maybe RequestCharged)

Undocumented member.

writeGetObjectResponse_eTag :: Lens' WriteGetObjectResponse (Maybe ETag)

An opaque identifier assigned by a web server to a specific version of a resource found at a URL.

writeGetObjectResponse_tagCount :: Lens' WriteGetObjectResponse (Maybe Int)

The number of tags, if any, on the object.

writeGetObjectResponse_contentType :: Lens' WriteGetObjectResponse (Maybe Text)

A standard MIME type describing the format of the object data.

writeGetObjectResponse_contentDisposition :: Lens' WriteGetObjectResponse (Maybe Text)

Specifies presentational information for the object.

writeGetObjectResponse_contentRange :: Lens' WriteGetObjectResponse (Maybe Text)

The portion of the object returned in the response.

writeGetObjectResponse_deleteMarker :: Lens' WriteGetObjectResponse (Maybe Bool)

Specifies whether an object stored in Amazon S3 is (true) or is not (false) a delete marker.

writeGetObjectResponse_contentLanguage :: Lens' WriteGetObjectResponse (Maybe Text)

The language the content is in.

writeGetObjectResponse_expiration :: Lens' WriteGetObjectResponse (Maybe Text)

If object stored in Amazon S3 expiration is configured (see PUT Bucket lifecycle) it includes expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL encoded.

writeGetObjectResponse_replicationStatus :: Lens' WriteGetObjectResponse (Maybe ReplicationStatus)

Indicates if request involves bucket that is either a source or destination in a Replication rule. For more information about S3 Replication, see Replication.

writeGetObjectResponse_contentLength :: Lens' WriteGetObjectResponse (Maybe Integer)

The size of the content body in bytes.

writeGetObjectResponse_metadata :: Lens' WriteGetObjectResponse (HashMap Text Text)

A map of metadata to store with the object in S3.

writeGetObjectResponse_sSECustomerKeyMD5 :: Lens' WriteGetObjectResponse (Maybe Text)

128-bit MD5 digest of customer-provided encryption key used in Amazon S3 to encrypt data stored in S3. For more information, see Protecting data using server-side encryption with customer-provided encryption keys (SSE-C).

writeGetObjectResponse_sSEKMSKeyId :: Lens' WriteGetObjectResponse (Maybe Text)

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for stored in Amazon S3 object.

writeGetObjectResponse_contentEncoding :: Lens' WriteGetObjectResponse (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

writeGetObjectResponse_versionId :: Lens' WriteGetObjectResponse (Maybe ObjectVersionId)

An ID used to reference a specific version of the object.

writeGetObjectResponse_storageClass :: Lens' WriteGetObjectResponse (Maybe StorageClass)

The class of storage used to store object in Amazon S3.

writeGetObjectResponse_bucketKeyEnabled :: Lens' WriteGetObjectResponse (Maybe Bool)

Indicates whether the object stored in Amazon S3 uses an S3 bucket key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

writeGetObjectResponse_acceptRanges :: Lens' WriteGetObjectResponse (Maybe Text)

Indicates that a range of bytes was specified.

writeGetObjectResponse_missingMeta :: Lens' WriteGetObjectResponse (Maybe Int)

Set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

writeGetObjectResponse_statusCode :: Lens' WriteGetObjectResponse (Maybe Int)

The integer status code for an HTTP response of a corresponding GetObject request.

Status Codes

200 - OK

206 - Partial Content

304 - Not Modified

400 - Bad Request

401 - Unauthorized

403 - Forbidden

404 - Not Found

405 - Method Not Allowed

409 - Conflict

411 - Length Required

412 - Precondition Failed

416 - Range Not Satisfiable

500 - Internal Server Error

503 - Service Unavailable

writeGetObjectResponse_serverSideEncryption :: Lens' WriteGetObjectResponse (Maybe ServerSideEncryption)

The server-side encryption algorithm used when storing requested object in Amazon S3 (for example, AES256, aws:kms).

writeGetObjectResponse_sSECustomerAlgorithm :: Lens' WriteGetObjectResponse (Maybe Text)

Encryption algorithm used if server-side encryption with a customer-provided encryption key was specified for object stored in Amazon S3.

writeGetObjectResponse_objectLockLegalHoldStatus :: Lens' WriteGetObjectResponse (Maybe ObjectLockLegalHoldStatus)

Indicates whether an object stored in Amazon S3 has an active legal hold.

writeGetObjectResponse_cacheControl :: Lens' WriteGetObjectResponse (Maybe Text)

Specifies caching behavior along the request/reply chain.

writeGetObjectResponse_errorMessage :: Lens' WriteGetObjectResponse (Maybe Text)

Contains a generic description of the error condition. Returned in the <Message> tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in body.

writeGetObjectResponse_lastModified :: Lens' WriteGetObjectResponse (Maybe UTCTime)

The date and time that the object was last modified.

writeGetObjectResponse_expires :: Lens' WriteGetObjectResponse (Maybe UTCTime)

The date and time at which the object is no longer cacheable.

writeGetObjectResponse_restore :: Lens' WriteGetObjectResponse (Maybe Text)

Provides information about object restoration operation and expiration time of the restored object copy.

writeGetObjectResponse_objectLockMode :: Lens' WriteGetObjectResponse (Maybe ObjectLockMode)

Indicates whether an object stored in Amazon S3 has Object Lock enabled. For more information about S3 Object Lock, see Object Lock.

writeGetObjectResponse_errorCode :: Lens' WriteGetObjectResponse (Maybe Text)

A string that uniquely identifies an error condition. Returned in the <Code> tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in the body. All error codes from S3 are sentence-cased. Regex value is "^[A-Z][a-zA-Z]+$".

writeGetObjectResponse_objectLockRetainUntilDate :: Lens' WriteGetObjectResponse (Maybe UTCTime)

The date and time when Object Lock is configured to expire.

writeGetObjectResponse_requestRoute :: Lens' WriteGetObjectResponse Text

Route prefix to the HTTP URL generated.

writeGetObjectResponse_requestToken :: Lens' WriteGetObjectResponse Text

A single use encrypted token that maps WriteGetObjectResponse to the end user GetObject request.

writeGetObjectResponse_body :: Lens' WriteGetObjectResponse RequestBody

The object data.

Destructuring the Response

data WriteGetObjectResponseResponse

See: newWriteGetObjectResponseResponse smart constructor.

Constructors

WriteGetObjectResponseResponse'

newWriteGetObjectResponseResponse :: WriteGetObjectResponseResponse

Create a value of WriteGetObjectResponseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.