Network.AWS.S3Outposts.CreateEndpoint

Description

Amazon S3 on Outposts Access Points simplify managing data access at scale for shared datasets in S3 on Outposts. S3 on Outposts uses endpoints to connect to Outposts buckets so that you can perform actions within your virtual private cloud (VPC). For more information, see Accessing S3 on Outposts using VPC only access points.

This action creates an endpoint and associates it with the specified Outposts.

It can take up to 5 minutes for this action to complete.

Related actions include:

DeleteEndpoint

ListEndpoints

Creating a Request

data CreateEndpoint

See: newCreateEndpoint smart constructor.

Constructors

CreateEndpoint'

Fields

accessType :: Maybe EndpointAccessType

The type of access for the on-premise network connectivity for the Outpost endpoint. To access the endpoint from an on-premises network, you must specify the access type and provide the customer owned IPv4 pool.

customerOwnedIpv4Pool :: Maybe Text

The ID of the customer-owned IPv4 pool for the endpoint. IP addresses will be allocated from this pool for the endpoint.

outpostId :: Text

The ID of the AWS Outposts.

subnetId :: Text

The ID of the subnet in the selected VPC. The endpoint subnet must belong to the Outpost that has the Amazon S3 on Outposts provisioned.

securityGroupId :: Text

The ID of the security group to use with the endpoint.

newCreateEndpoint

Arguments

:: Text
$sel:outpostId:CreateEndpoint'

-> Text
$sel:subnetId:CreateEndpoint'

-> Text
$sel:securityGroupId:CreateEndpoint'

-> CreateEndpoint

Create a value of CreateEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessType:CreateEndpoint', createEndpoint_accessType - The type of access for the on-premise network connectivity for the Outpost endpoint. To access the endpoint from an on-premises network, you must specify the access type and provide the customer owned IPv4 pool.

$sel:customerOwnedIpv4Pool:CreateEndpoint', createEndpoint_customerOwnedIpv4Pool - The ID of the customer-owned IPv4 pool for the endpoint. IP addresses will be allocated from this pool for the endpoint.

$sel:outpostId:CreateEndpoint', createEndpoint_outpostId - The ID of the AWS Outposts.

$sel:subnetId:CreateEndpoint', createEndpoint_subnetId - The ID of the subnet in the selected VPC. The endpoint subnet must belong to the Outpost that has the Amazon S3 on Outposts provisioned.

$sel:securityGroupId:CreateEndpoint', createEndpoint_securityGroupId - The ID of the security group to use with the endpoint.

Request Lenses

createEndpoint_accessType :: Lens' CreateEndpoint (Maybe EndpointAccessType)

The type of access for the on-premise network connectivity for the Outpost endpoint. To access the endpoint from an on-premises network, you must specify the access type and provide the customer owned IPv4 pool.

createEndpoint_customerOwnedIpv4Pool :: Lens' CreateEndpoint (Maybe Text)

The ID of the customer-owned IPv4 pool for the endpoint. IP addresses will be allocated from this pool for the endpoint.

createEndpoint_outpostId :: Lens' CreateEndpoint Text

The ID of the AWS Outposts.

createEndpoint_subnetId :: Lens' CreateEndpoint Text

The ID of the subnet in the selected VPC. The endpoint subnet must belong to the Outpost that has the Amazon S3 on Outposts provisioned.

createEndpoint_securityGroupId :: Lens' CreateEndpoint Text

The ID of the security group to use with the endpoint.

Destructuring the Response

data CreateEndpointResponse

See: newCreateEndpointResponse smart constructor.

Constructors

CreateEndpointResponse'

Fields

endpointArn :: Maybe Text

The Amazon Resource Name (ARN) of the endpoint.

httpStatus :: Int

The response's http status code.

newCreateEndpointResponse

Arguments

:: Int
$sel:httpStatus:CreateEndpointResponse'

-> CreateEndpointResponse

Create a value of CreateEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpointArn:CreateEndpointResponse', createEndpointResponse_endpointArn - The Amazon Resource Name (ARN) of the endpoint.

$sel:httpStatus:CreateEndpointResponse', createEndpointResponse_httpStatus - The response's http status code.

Response Lenses

createEndpointResponse_endpointArn :: Lens' CreateEndpointResponse (Maybe Text)

The Amazon Resource Name (ARN) of the endpoint.

createEndpointResponse_httpStatus :: Lens' CreateEndpointResponse Int

The response's http status code.