Network.AWS.S3Outposts.Types

Description

Service Configuration

defaultService :: Service

API version 2017-07-25 of the Amazon S3 on Outposts SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

There was an exception validating this data.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

Access was denied for this action.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

There was a conflict with this action, and it could not be completed.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

There was an exception with the internal server.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The requested resource was not found.

EndpointAccessType

newtype EndpointAccessType

Constructors

EndpointAccessType'

Fields

fromEndpointAccessType :: Text

Bundled Patterns

pattern EndpointAccessType_CustomerOwnedIp :: EndpointAccessType

pattern EndpointAccessType_Private :: EndpointAccessType

EndpointStatus

newtype EndpointStatus

Constructors

EndpointStatus'

Fields

fromEndpointStatus :: Text

Bundled Patterns

pattern EndpointStatus_Available :: EndpointStatus

pattern EndpointStatus_Deleting :: EndpointStatus

pattern EndpointStatus_Pending :: EndpointStatus

Endpoint

data Endpoint

Amazon S3 on Outposts Access Points simplify managing data access at scale for shared datasets in S3 on Outposts. S3 on Outposts uses endpoints to connect to Outposts buckets so that you can perform actions within your virtual private cloud (VPC). For more information, see Accessing S3 on Outposts using VPC only access points.

See: newEndpoint smart constructor.

Constructors

Endpoint'

Fields

creationTime :: Maybe POSIX

The time the endpoint was created.

status :: Maybe EndpointStatus

The status of the endpoint.

vpcId :: Maybe Text

The ID of the VPC used for the endpoint.

networkInterfaces :: Maybe [NetworkInterface]

The network interface of the endpoint.

outpostsId :: Maybe Text

The ID of the AWS Outposts.

subnetId :: Maybe Text

The ID of the subnet used for the endpoint.

securityGroupId :: Maybe Text

The ID of the security group used for the endpoint.

accessType :: Maybe EndpointAccessType

customerOwnedIpv4Pool :: Maybe Text

The ID of the customer-owned IPv4 pool used for the endpoint.

cidrBlock :: Maybe Text

The VPC CIDR committed by this endpoint.

endpointArn :: Maybe Text

The Amazon Resource Name (ARN) of the endpoint.

newEndpoint :: Endpoint

Create a value of Endpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Endpoint', endpoint_creationTime - The time the endpoint was created.

$sel:status:Endpoint', endpoint_status - The status of the endpoint.

$sel:vpcId:Endpoint', endpoint_vpcId - The ID of the VPC used for the endpoint.

$sel:networkInterfaces:Endpoint', endpoint_networkInterfaces - The network interface of the endpoint.

$sel:outpostsId:Endpoint', endpoint_outpostsId - The ID of the AWS Outposts.

$sel:subnetId:Endpoint', endpoint_subnetId - The ID of the subnet used for the endpoint.

$sel:securityGroupId:Endpoint', endpoint_securityGroupId - The ID of the security group used for the endpoint.

$sel:accessType:Endpoint', endpoint_accessType -

$sel:customerOwnedIpv4Pool:Endpoint', endpoint_customerOwnedIpv4Pool - The ID of the customer-owned IPv4 pool used for the endpoint.

$sel:cidrBlock:Endpoint', endpoint_cidrBlock - The VPC CIDR committed by this endpoint.

$sel:endpointArn:Endpoint', endpoint_endpointArn - The Amazon Resource Name (ARN) of the endpoint.

endpoint_creationTime :: Lens' Endpoint (Maybe UTCTime)

The time the endpoint was created.

endpoint_status :: Lens' Endpoint (Maybe EndpointStatus)

The status of the endpoint.

endpoint_vpcId :: Lens' Endpoint (Maybe Text)

The ID of the VPC used for the endpoint.

endpoint_networkInterfaces :: Lens' Endpoint (Maybe [NetworkInterface])

The network interface of the endpoint.

endpoint_outpostsId :: Lens' Endpoint (Maybe Text)

The ID of the AWS Outposts.

endpoint_subnetId :: Lens' Endpoint (Maybe Text)

The ID of the subnet used for the endpoint.

endpoint_securityGroupId :: Lens' Endpoint (Maybe Text)

The ID of the security group used for the endpoint.

endpoint_accessType :: Lens' Endpoint (Maybe EndpointAccessType)

endpoint_customerOwnedIpv4Pool :: Lens' Endpoint (Maybe Text)

The ID of the customer-owned IPv4 pool used for the endpoint.

endpoint_cidrBlock :: Lens' Endpoint (Maybe Text)

The VPC CIDR committed by this endpoint.

endpoint_endpointArn :: Lens' Endpoint (Maybe Text)

The Amazon Resource Name (ARN) of the endpoint.

NetworkInterface

data NetworkInterface

The container for the network interface.

See: newNetworkInterface smart constructor.

Constructors

NetworkInterface'

Fields

networkInterfaceId :: Maybe Text

The ID for the network interface.

newNetworkInterface :: NetworkInterface

Create a value of NetworkInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkInterfaceId:NetworkInterface', networkInterface_networkInterfaceId - The ID for the network interface.

networkInterface_networkInterfaceId :: Lens' NetworkInterface (Maybe Text)

The ID for the network interface.