Network.AWS.SDB

Description

Derived from API version 2009-04-15 of the AWS service descriptions, licensed under Apache 2.0.

Amazon SimpleDB is a web service providing the core database functions of data indexing and querying in the cloud. By offloading the time and effort associated with building and operating a web-scale database, SimpleDB provides developers the freedom to focus on application development.

A traditional, clustered relational database requires a sizable upfront capital outlay, is complex to design, and often requires extensive and repetitive database administration. Amazon SimpleDB is dramatically simpler, requiring no schema, automatically indexing your data and providing a simple API for storage and access. This approach eliminates the administrative burden of data modeling, index maintenance, and performance tuning. Developers gain access to this functionality within Amazon's proven computing environment, are able to scale instantly, and pay only for what they use.

Visit http://aws.amazon.com/simpledb/ for more information.

Service Configuration

defaultService :: Service

API version 2009-04-15 of the Amazon SimpleDB SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by SDB.

NumberDomainAttributesExceeded

_NumberDomainAttributesExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Too many attributes in this domain.

NumberSubmittedItemsExceeded

_NumberSubmittedItemsExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Too many items exist in a single call.

NoSuchDomain

_NoSuchDomain :: AsError a => Getting (First ServiceError) a ServiceError

The specified domain does not exist.

NumberDomainBytesExceeded

_NumberDomainBytesExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Too many bytes in this domain.

NumberDomainsExceeded

_NumberDomainsExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Too many domains exist per this account.

InvalidNumberPredicates

_InvalidNumberPredicates :: AsError a => Getting (First ServiceError) a ServiceError

Too many predicates exist in the query expression.

DuplicateItemName

_DuplicateItemName :: AsError a => Getting (First ServiceError) a ServiceError

The item name was specified more than once.

AttributeDoesNotExist

_AttributeDoesNotExist :: AsError a => Getting (First ServiceError) a ServiceError

The specified attribute does not exist.

InvalidNumberValueTests

_InvalidNumberValueTests :: AsError a => Getting (First ServiceError) a ServiceError

Too many predicates exist in the query expression.

InvalidQueryExpression

_InvalidQueryExpression :: AsError a => Getting (First ServiceError) a ServiceError

The specified query expression syntax is not valid.

NumberSubmittedAttributesExceeded

_NumberSubmittedAttributesExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Too many attributes exist in a single call.

InvalidParameterValue

_InvalidParameterValue :: AsError a => Getting (First ServiceError) a ServiceError

The value for a parameter is invalid.

TooManyRequestedAttributes

_TooManyRequestedAttributes :: AsError a => Getting (First ServiceError) a ServiceError

Too many attributes requested.

InvalidNextToken

_InvalidNextToken :: AsError a => Getting (First ServiceError) a ServiceError

The specified NextToken is not valid.

MissingParameter

_MissingParameter :: AsError a => Getting (First ServiceError) a ServiceError

The request must contain the specified missing parameter.

NumberItemAttributesExceeded

_NumberItemAttributesExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Too many attributes in this item.

RequestTimeout

_RequestTimeout :: AsError a => Getting (First ServiceError) a ServiceError

A timeout occurred when attempting to query the specified domain with specified query expression.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListDomains (Paginated)

data ListDomains

See: newListDomains smart constructor.

Constructors

ListDomains' (Maybe Text) (Maybe Int)

newListDomains :: ListDomains

Create a value of ListDomains with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDomains', listDomains_nextToken - A string informing Amazon SimpleDB where to start the next list of domain names.

$sel:maxNumberOfDomains:ListDomains', listDomains_maxNumberOfDomains - The maximum number of domain names you want returned. The range is 1 to 100. The default setting is 100.

data ListDomainsResponse

See: newListDomainsResponse smart constructor.

Constructors

ListDomainsResponse' (Maybe [Text]) (Maybe Text) Int

newListDomainsResponse

Arguments

:: Int
$sel:httpStatus:ListDomainsResponse'

-> ListDomainsResponse

Create a value of ListDomainsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainNames:ListDomainsResponse', listDomainsResponse_domainNames - A list of domain names that match the expression.

$sel:nextToken:ListDomains', listDomainsResponse_nextToken - An opaque token indicating that there are more domains than the specified MaxNumberOfDomains still available.

$sel:httpStatus:ListDomainsResponse', listDomainsResponse_httpStatus - The response's http status code.

DeleteAttributes

data DeleteAttributes

See: newDeleteAttributes smart constructor.

Constructors

DeleteAttributes' (Maybe UpdateCondition) (Maybe [Attribute]) Text Text

newDeleteAttributes

Arguments

:: Text
$sel:domainName:DeleteAttributes'

-> Text
$sel:itemName:DeleteAttributes'

-> DeleteAttributes

Create a value of DeleteAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expected:DeleteAttributes', deleteAttributes_expected - The update condition which, if specified, determines whether the specified attributes will be deleted or not. The update condition must be satisfied in order for this request to be processed and the attributes to be deleted.

$sel:attributes:DeleteAttributes', deleteAttributes_attributes - A list of Attributes. Similar to columns on a spreadsheet, attributes represent categories of data that can be assigned to items.

$sel:domainName:DeleteAttributes', deleteAttributes_domainName - The name of the domain in which to perform the operation.

$sel:itemName:DeleteAttributes', deleteAttributes_itemName - The name of the item. Similar to rows on a spreadsheet, items represent individual objects that contain one or more value-attribute pairs.

data DeleteAttributesResponse

See: newDeleteAttributesResponse smart constructor.

Constructors

DeleteAttributesResponse'

newDeleteAttributesResponse :: DeleteAttributesResponse

Create a value of DeleteAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

DomainMetadata

data DomainMetadata

See: newDomainMetadata smart constructor.

Constructors

DomainMetadata' Text

newDomainMetadata

Arguments

:: Text
$sel:domainName:DomainMetadata'

-> DomainMetadata

Create a value of DomainMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:DomainMetadata', domainMetadata_domainName - The name of the domain for which to display the metadata of.

data DomainMetadataResponse

See: newDomainMetadataResponse smart constructor.

Constructors

DomainMetadataResponse' (Maybe Integer) (Maybe Int) (Maybe Int) (Maybe Integer) (Maybe Int) (Maybe Int) (Maybe Integer) Int

newDomainMetadataResponse

Arguments

:: Int
$sel:httpStatus:DomainMetadataResponse'

-> DomainMetadataResponse

Create a value of DomainMetadataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeNamesSizeBytes:DomainMetadataResponse', domainMetadataResponse_attributeNamesSizeBytes - The total size of all unique attribute names in the domain, in bytes.

$sel:attributeValueCount:DomainMetadataResponse', domainMetadataResponse_attributeValueCount - The number of all attribute name/value pairs in the domain.

$sel:attributeNameCount:DomainMetadataResponse', domainMetadataResponse_attributeNameCount - The number of unique attribute names in the domain.

$sel:attributeValuesSizeBytes:DomainMetadataResponse', domainMetadataResponse_attributeValuesSizeBytes - The total size of all attribute values in the domain, in bytes.

$sel:timestamp:DomainMetadataResponse', domainMetadataResponse_timestamp - The data and time when metadata was calculated, in Epoch (UNIX) seconds.

$sel:itemCount:DomainMetadataResponse', domainMetadataResponse_itemCount - The number of all items in the domain.

$sel:itemNamesSizeBytes:DomainMetadataResponse', domainMetadataResponse_itemNamesSizeBytes - The total size of all item names in the domain, in bytes.

$sel:httpStatus:DomainMetadataResponse', domainMetadataResponse_httpStatus - The response's http status code.

BatchPutAttributes

data BatchPutAttributes

See: newBatchPutAttributes smart constructor.

Constructors

BatchPutAttributes' Text [ReplaceableItem]

newBatchPutAttributes

Arguments

:: Text
$sel:domainName:BatchPutAttributes'

-> BatchPutAttributes

Create a value of BatchPutAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:BatchPutAttributes', batchPutAttributes_domainName - The name of the domain in which the attributes are being stored.

$sel:items:BatchPutAttributes', batchPutAttributes_items - A list of items on which to perform the operation.

data BatchPutAttributesResponse

See: newBatchPutAttributesResponse smart constructor.

Constructors

BatchPutAttributesResponse'

newBatchPutAttributesResponse :: BatchPutAttributesResponse

Create a value of BatchPutAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

CreateDomain

data CreateDomain

See: newCreateDomain smart constructor.

Constructors

CreateDomain' Text

newCreateDomain

Arguments

:: Text
$sel:domainName:CreateDomain'

-> CreateDomain

Create a value of CreateDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:CreateDomain', createDomain_domainName - The name of the domain to create. The name can range between 3 and 255 characters and can contain the following characters: a-z, A-Z, 0-9, '_', '-', and '.'.

data CreateDomainResponse

See: newCreateDomainResponse smart constructor.

Constructors

CreateDomainResponse'

newCreateDomainResponse :: CreateDomainResponse

Create a value of CreateDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

DeleteDomain

data DeleteDomain

See: newDeleteDomain smart constructor.

Constructors

DeleteDomain' Text

newDeleteDomain

Arguments

:: Text
$sel:domainName:DeleteDomain'

-> DeleteDomain

Create a value of DeleteDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:DeleteDomain', deleteDomain_domainName - The name of the domain to delete.

data DeleteDomainResponse

See: newDeleteDomainResponse smart constructor.

Constructors

DeleteDomainResponse'

newDeleteDomainResponse :: DeleteDomainResponse

Create a value of DeleteDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

GetAttributes

data GetAttributes

See: newGetAttributes smart constructor.

Constructors

GetAttributes' (Maybe Bool) (Maybe [Text]) Text Text

newGetAttributes

Arguments

:: Text
$sel:domainName:GetAttributes'

-> Text
$sel:itemName:GetAttributes'

-> GetAttributes

Create a value of GetAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:consistentRead:GetAttributes', getAttributes_consistentRead - Determines whether or not strong consistency should be enforced when data is read from SimpleDB. If true, any data previously written to SimpleDB will be returned. Otherwise, results will be consistent eventually, and the client may not see data that was written immediately before your read.

$sel:attributeNames:GetAttributes', getAttributes_attributeNames - The names of the attributes.

$sel:domainName:GetAttributes', getAttributes_domainName - The name of the domain in which to perform the operation.

$sel:itemName:GetAttributes', getAttributes_itemName - The name of the item.

data GetAttributesResponse

See: newGetAttributesResponse smart constructor.

Constructors

GetAttributesResponse' (Maybe [Attribute]) Int

newGetAttributesResponse

Arguments

:: Int
$sel:httpStatus:GetAttributesResponse'

-> GetAttributesResponse

Create a value of GetAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:GetAttributesResponse', getAttributesResponse_attributes - The list of attributes returned by the operation.

$sel:httpStatus:GetAttributesResponse', getAttributesResponse_httpStatus - The response's http status code.

PutAttributes

data PutAttributes

See: newPutAttributes smart constructor.

Constructors

PutAttributes' (Maybe UpdateCondition) Text Text [ReplaceableAttribute]

newPutAttributes

Arguments

:: Text
$sel:domainName:PutAttributes'

-> Text
$sel:itemName:PutAttributes'

-> PutAttributes

Create a value of PutAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expected:PutAttributes', putAttributes_expected - The update condition which, if specified, determines whether the specified attributes will be updated or not. The update condition must be satisfied in order for this request to be processed and the attributes to be updated.

$sel:domainName:PutAttributes', putAttributes_domainName - The name of the domain in which to perform the operation.

$sel:itemName:PutAttributes', putAttributes_itemName - The name of the item.

$sel:attributes:PutAttributes', putAttributes_attributes - The list of attributes.

data PutAttributesResponse

See: newPutAttributesResponse smart constructor.

Constructors

PutAttributesResponse'

newPutAttributesResponse :: PutAttributesResponse

Create a value of PutAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Select (Paginated)

data Select

See: newSelect smart constructor.

Constructors

Select' (Maybe Text) (Maybe Bool) Text

newSelect

Arguments

:: Text
$sel:selectExpression:Select'

-> Select

Create a value of Select with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:Select', select_nextToken - A string informing Amazon SimpleDB where to start the next list of ItemNames.

$sel:consistentRead:Select', select_consistentRead - Determines whether or not strong consistency should be enforced when data is read from SimpleDB. If true, any data previously written to SimpleDB will be returned. Otherwise, results will be consistent eventually, and the client may not see data that was written immediately before your read.

$sel:selectExpression:Select', select_selectExpression - The expression used to query the domain.

data SelectResponse

See: newSelectResponse smart constructor.

Constructors

SelectResponse' (Maybe Text) (Maybe [Item]) Int

newSelectResponse

Arguments

:: Int
$sel:httpStatus:SelectResponse'

-> SelectResponse

Create a value of SelectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:Select', selectResponse_nextToken - An opaque token indicating that more items than MaxNumberOfItems were matched, the response size exceeded 1 megabyte, or the execution time exceeded 5 seconds.

$sel:items:SelectResponse', selectResponse_items - A list of items that match the select expression.

$sel:httpStatus:SelectResponse', selectResponse_httpStatus - The response's http status code.

BatchDeleteAttributes

data BatchDeleteAttributes

See: newBatchDeleteAttributes smart constructor.

Constructors

BatchDeleteAttributes' Text [DeletableItem]

newBatchDeleteAttributes

Arguments

:: Text
$sel:domainName:BatchDeleteAttributes'

-> BatchDeleteAttributes

Create a value of BatchDeleteAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:BatchDeleteAttributes', batchDeleteAttributes_domainName - The name of the domain in which the attributes are being deleted.

$sel:items:BatchDeleteAttributes', batchDeleteAttributes_items - A list of items on which to perform the operation.

data BatchDeleteAttributesResponse

See: newBatchDeleteAttributesResponse smart constructor.

Constructors

BatchDeleteAttributesResponse'

newBatchDeleteAttributesResponse :: BatchDeleteAttributesResponse

Create a value of BatchDeleteAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Types

Attribute

data Attribute

See: newAttribute smart constructor.

Constructors

Attribute' (Maybe Text) (Maybe Text) Text Text

newAttribute

Arguments

:: Text
$sel:name:Attribute'

-> Text
$sel:value:Attribute'

-> Attribute

Create a value of Attribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alternateNameEncoding:Attribute', attribute_alternateNameEncoding -

$sel:alternateValueEncoding:Attribute', attribute_alternateValueEncoding -

$sel:name:Attribute', attribute_name - The name of the attribute.

$sel:value:Attribute', attribute_value - The value of the attribute.

DeletableItem

data DeletableItem

See: newDeletableItem smart constructor.

Constructors

DeletableItem' (Maybe [Attribute]) Text

newDeletableItem

Arguments

:: Text
$sel:name:DeletableItem'

-> DeletableItem

Create a value of DeletableItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:DeletableItem', deletableItem_attributes - Undocumented member.

$sel:name:DeletableItem', deletableItem_name - Undocumented member.

Item

data Item

See: newItem smart constructor.

Constructors

Item' (Maybe Text) Text [Attribute]

newItem

Arguments

:: Text
$sel:name:Item'

-> Item

Create a value of Item with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alternateNameEncoding:Item', item_alternateNameEncoding -

$sel:name:Item', item_name - The name of the item.

$sel:attributes:Item', item_attributes - A list of attributes.

ReplaceableAttribute

data ReplaceableAttribute

See: newReplaceableAttribute smart constructor.

Constructors

ReplaceableAttribute' (Maybe Bool) Text Text

newReplaceableAttribute

Arguments

:: Text
$sel:name:ReplaceableAttribute'

-> Text
$sel:value:ReplaceableAttribute'

-> ReplaceableAttribute

Create a value of ReplaceableAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replace:ReplaceableAttribute', replaceableAttribute_replace - A flag specifying whether or not to replace the attribute/value pair or to add a new attribute/value pair. The default setting is false.

$sel:name:ReplaceableAttribute', replaceableAttribute_name - The name of the replaceable attribute.

$sel:value:ReplaceableAttribute', replaceableAttribute_value - The value of the replaceable attribute.

ReplaceableItem

data ReplaceableItem

See: newReplaceableItem smart constructor.

Constructors

ReplaceableItem' Text [ReplaceableAttribute]

newReplaceableItem

Arguments

:: Text
$sel:name:ReplaceableItem'

-> ReplaceableItem

Create a value of ReplaceableItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ReplaceableItem', replaceableItem_name - The name of the replaceable item.

$sel:attributes:ReplaceableItem', replaceableItem_attributes - The list of attributes for a replaceable item.

UpdateCondition

data UpdateCondition

Specifies the conditions under which data should be updated. If an update condition is specified for a request, the data will only be updated if the condition is satisfied. For example, if an attribute with a specific name and value exists, or if a specific attribute doesn't exist.

See: newUpdateCondition smart constructor.

Constructors

UpdateCondition' (Maybe Bool) (Maybe Text) (Maybe Text)

newUpdateCondition :: UpdateCondition

Create a value of UpdateCondition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exists:UpdateCondition', updateCondition_exists - A value specifying whether or not the specified attribute must exist with the specified value in order for the update condition to be satisfied. Specify true if the attribute must exist for the update condition to be satisfied. Specify false if the attribute should not exist in order for the update condition to be satisfied.

$sel:name:UpdateCondition', updateCondition_name - The name of the attribute involved in the condition.

$sel:value:UpdateCondition', updateCondition_value - The value of an attribute. This value can only be specified when the Exists parameter is equal to true.