Network.AWS.SDB.DeleteAttributes

Description

Deletes one or more attributes associated with an item. If all attributes of the item are deleted, the item is deleted.

DeleteAttributes is an idempotent operation; running it multiple times on the same item or attribute does not result in an error response.

Because Amazon SimpleDB makes multiple copies of item data and uses an eventual consistency update model, performing a GetAttributes or Select operation (read) immediately after a DeleteAttributes or PutAttributes operation (write) might not return updated item data.

Creating a Request

data DeleteAttributes

See: newDeleteAttributes smart constructor.

Constructors

DeleteAttributes'

Fields

expected :: Maybe UpdateCondition

The update condition which, if specified, determines whether the specified attributes will be deleted or not. The update condition must be satisfied in order for this request to be processed and the attributes to be deleted.

attributes :: Maybe [Attribute]

A list of Attributes. Similar to columns on a spreadsheet, attributes represent categories of data that can be assigned to items.

domainName :: Text

The name of the domain in which to perform the operation.

itemName :: Text

The name of the item. Similar to rows on a spreadsheet, items represent individual objects that contain one or more value-attribute pairs.

newDeleteAttributes

Arguments

:: Text
$sel:domainName:DeleteAttributes'

-> Text
$sel:itemName:DeleteAttributes'

-> DeleteAttributes

Create a value of DeleteAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expected:DeleteAttributes', deleteAttributes_expected - The update condition which, if specified, determines whether the specified attributes will be deleted or not. The update condition must be satisfied in order for this request to be processed and the attributes to be deleted.

$sel:attributes:DeleteAttributes', deleteAttributes_attributes - A list of Attributes. Similar to columns on a spreadsheet, attributes represent categories of data that can be assigned to items.

$sel:domainName:DeleteAttributes', deleteAttributes_domainName - The name of the domain in which to perform the operation.

$sel:itemName:DeleteAttributes', deleteAttributes_itemName - The name of the item. Similar to rows on a spreadsheet, items represent individual objects that contain one or more value-attribute pairs.

Request Lenses

deleteAttributes_expected :: Lens' DeleteAttributes (Maybe UpdateCondition)

The update condition which, if specified, determines whether the specified attributes will be deleted or not. The update condition must be satisfied in order for this request to be processed and the attributes to be deleted.

deleteAttributes_attributes :: Lens' DeleteAttributes (Maybe [Attribute])

A list of Attributes. Similar to columns on a spreadsheet, attributes represent categories of data that can be assigned to items.

deleteAttributes_domainName :: Lens' DeleteAttributes Text

The name of the domain in which to perform the operation.

deleteAttributes_itemName :: Lens' DeleteAttributes Text

The name of the item. Similar to rows on a spreadsheet, items represent individual objects that contain one or more value-attribute pairs.

Destructuring the Response

data DeleteAttributesResponse

See: newDeleteAttributesResponse smart constructor.

Constructors

DeleteAttributesResponse'

newDeleteAttributesResponse :: DeleteAttributesResponse

Create a value of DeleteAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.