Network.AWS.SDB.DomainMetadata

Description

Returns information about the domain, including when the domain was created, the number of items and attributes in the domain, and the size of the attribute names and values.

Creating a Request

data DomainMetadata

See: newDomainMetadata smart constructor.

Constructors

DomainMetadata'

Fields

domainName :: Text

The name of the domain for which to display the metadata of.

newDomainMetadata

Arguments

:: Text
$sel:domainName:DomainMetadata'

-> DomainMetadata

Create a value of DomainMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:DomainMetadata', domainMetadata_domainName - The name of the domain for which to display the metadata of.

Request Lenses

domainMetadata_domainName :: Lens' DomainMetadata Text

The name of the domain for which to display the metadata of.

Destructuring the Response

data DomainMetadataResponse

See: newDomainMetadataResponse smart constructor.

Constructors

DomainMetadataResponse'

Fields

attributeNamesSizeBytes :: Maybe Integer

The total size of all unique attribute names in the domain, in bytes.

attributeValueCount :: Maybe Int

The number of all attribute name/value pairs in the domain.

attributeNameCount :: Maybe Int

The number of unique attribute names in the domain.

attributeValuesSizeBytes :: Maybe Integer

The total size of all attribute values in the domain, in bytes.

timestamp :: Maybe Int

The data and time when metadata was calculated, in Epoch (UNIX) seconds.

itemCount :: Maybe Int

The number of all items in the domain.

itemNamesSizeBytes :: Maybe Integer

The total size of all item names in the domain, in bytes.

httpStatus :: Int

The response's http status code.

newDomainMetadataResponse

Arguments

:: Int
$sel:httpStatus:DomainMetadataResponse'

-> DomainMetadataResponse

Create a value of DomainMetadataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeNamesSizeBytes:DomainMetadataResponse', domainMetadataResponse_attributeNamesSizeBytes - The total size of all unique attribute names in the domain, in bytes.

$sel:attributeValueCount:DomainMetadataResponse', domainMetadataResponse_attributeValueCount - The number of all attribute name/value pairs in the domain.

$sel:attributeNameCount:DomainMetadataResponse', domainMetadataResponse_attributeNameCount - The number of unique attribute names in the domain.

$sel:attributeValuesSizeBytes:DomainMetadataResponse', domainMetadataResponse_attributeValuesSizeBytes - The total size of all attribute values in the domain, in bytes.

$sel:timestamp:DomainMetadataResponse', domainMetadataResponse_timestamp - The data and time when metadata was calculated, in Epoch (UNIX) seconds.

$sel:itemCount:DomainMetadataResponse', domainMetadataResponse_itemCount - The number of all items in the domain.

$sel:itemNamesSizeBytes:DomainMetadataResponse', domainMetadataResponse_itemNamesSizeBytes - The total size of all item names in the domain, in bytes.

$sel:httpStatus:DomainMetadataResponse', domainMetadataResponse_httpStatus - The response's http status code.

Response Lenses

domainMetadataResponse_attributeNamesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

The total size of all unique attribute names in the domain, in bytes.

domainMetadataResponse_attributeValueCount :: Lens' DomainMetadataResponse (Maybe Int)

The number of all attribute name/value pairs in the domain.

domainMetadataResponse_attributeNameCount :: Lens' DomainMetadataResponse (Maybe Int)

The number of unique attribute names in the domain.

domainMetadataResponse_attributeValuesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

The total size of all attribute values in the domain, in bytes.

domainMetadataResponse_timestamp :: Lens' DomainMetadataResponse (Maybe Int)

The data and time when metadata was calculated, in Epoch (UNIX) seconds.

domainMetadataResponse_itemCount :: Lens' DomainMetadataResponse (Maybe Int)

The number of all items in the domain.

domainMetadataResponse_itemNamesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

The total size of all item names in the domain, in bytes.

domainMetadataResponse_httpStatus :: Lens' DomainMetadataResponse Int

The response's http status code.