Network.AWS.SDB.GetAttributes

Description

Returns all of the attributes associated with the specified item. Optionally, the attributes returned can be limited to one or more attributes by specifying an attribute name parameter.

If the item does not exist on the replica that was accessed for this operation, an empty set is returned. The system does not return an error as it cannot guarantee the item does not exist on other replicas.

Creating a Request

data GetAttributes

See: newGetAttributes smart constructor.

Constructors

GetAttributes'

Fields

consistentRead :: Maybe Bool

Determines whether or not strong consistency should be enforced when data is read from SimpleDB. If true, any data previously written to SimpleDB will be returned. Otherwise, results will be consistent eventually, and the client may not see data that was written immediately before your read.

attributeNames :: Maybe [Text]

The names of the attributes.

domainName :: Text

The name of the domain in which to perform the operation.

itemName :: Text

The name of the item.

newGetAttributes

Arguments

:: Text
$sel:domainName:GetAttributes'

-> Text
$sel:itemName:GetAttributes'

-> GetAttributes

Create a value of GetAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:consistentRead:GetAttributes', getAttributes_consistentRead - Determines whether or not strong consistency should be enforced when data is read from SimpleDB. If true, any data previously written to SimpleDB will be returned. Otherwise, results will be consistent eventually, and the client may not see data that was written immediately before your read.

$sel:attributeNames:GetAttributes', getAttributes_attributeNames - The names of the attributes.

$sel:domainName:GetAttributes', getAttributes_domainName - The name of the domain in which to perform the operation.

$sel:itemName:GetAttributes', getAttributes_itemName - The name of the item.

Request Lenses

getAttributes_consistentRead :: Lens' GetAttributes (Maybe Bool)

Determines whether or not strong consistency should be enforced when data is read from SimpleDB. If true, any data previously written to SimpleDB will be returned. Otherwise, results will be consistent eventually, and the client may not see data that was written immediately before your read.

getAttributes_attributeNames :: Lens' GetAttributes (Maybe [Text])

The names of the attributes.

getAttributes_domainName :: Lens' GetAttributes Text

The name of the domain in which to perform the operation.

getAttributes_itemName :: Lens' GetAttributes Text

The name of the item.

Destructuring the Response

data GetAttributesResponse

See: newGetAttributesResponse smart constructor.

Constructors

GetAttributesResponse'

Fields

attributes :: Maybe [Attribute]

The list of attributes returned by the operation.

httpStatus :: Int

The response's http status code.

newGetAttributesResponse

Arguments

:: Int
$sel:httpStatus:GetAttributesResponse'

-> GetAttributesResponse

Create a value of GetAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:GetAttributesResponse', getAttributesResponse_attributes - The list of attributes returned by the operation.

$sel:httpStatus:GetAttributesResponse', getAttributesResponse_httpStatus - The response's http status code.

Response Lenses

getAttributesResponse_attributes :: Lens' GetAttributesResponse (Maybe [Attribute])

The list of attributes returned by the operation.

getAttributesResponse_httpStatus :: Lens' GetAttributesResponse Int

The response's http status code.