Network.AWS.SDB.Lens

Description

Operations

ListDomains

listDomains_nextToken :: Lens' ListDomains (Maybe Text)

A string informing Amazon SimpleDB where to start the next list of domain names.

listDomains_maxNumberOfDomains :: Lens' ListDomains (Maybe Int)

The maximum number of domain names you want returned. The range is 1 to 100. The default setting is 100.

listDomainsResponse_domainNames :: Lens' ListDomainsResponse (Maybe [Text])

A list of domain names that match the expression.

listDomainsResponse_nextToken :: Lens' ListDomainsResponse (Maybe Text)

An opaque token indicating that there are more domains than the specified MaxNumberOfDomains still available.

listDomainsResponse_httpStatus :: Lens' ListDomainsResponse Int

The response's http status code.

DeleteAttributes

deleteAttributes_expected :: Lens' DeleteAttributes (Maybe UpdateCondition)

The update condition which, if specified, determines whether the specified attributes will be deleted or not. The update condition must be satisfied in order for this request to be processed and the attributes to be deleted.

deleteAttributes_attributes :: Lens' DeleteAttributes (Maybe [Attribute])

A list of Attributes. Similar to columns on a spreadsheet, attributes represent categories of data that can be assigned to items.

deleteAttributes_domainName :: Lens' DeleteAttributes Text

The name of the domain in which to perform the operation.

deleteAttributes_itemName :: Lens' DeleteAttributes Text

The name of the item. Similar to rows on a spreadsheet, items represent individual objects that contain one or more value-attribute pairs.

DomainMetadata

domainMetadata_domainName :: Lens' DomainMetadata Text

The name of the domain for which to display the metadata of.

domainMetadataResponse_attributeNamesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

The total size of all unique attribute names in the domain, in bytes.

domainMetadataResponse_attributeValueCount :: Lens' DomainMetadataResponse (Maybe Int)

The number of all attribute name/value pairs in the domain.

domainMetadataResponse_attributeNameCount :: Lens' DomainMetadataResponse (Maybe Int)

The number of unique attribute names in the domain.

domainMetadataResponse_attributeValuesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

The total size of all attribute values in the domain, in bytes.

domainMetadataResponse_timestamp :: Lens' DomainMetadataResponse (Maybe Int)

The data and time when metadata was calculated, in Epoch (UNIX) seconds.

domainMetadataResponse_itemCount :: Lens' DomainMetadataResponse (Maybe Int)

The number of all items in the domain.

domainMetadataResponse_itemNamesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

The total size of all item names in the domain, in bytes.

domainMetadataResponse_httpStatus :: Lens' DomainMetadataResponse Int

The response's http status code.

BatchPutAttributes

batchPutAttributes_domainName :: Lens' BatchPutAttributes Text

The name of the domain in which the attributes are being stored.

batchPutAttributes_items :: Lens' BatchPutAttributes [ReplaceableItem]

A list of items on which to perform the operation.

CreateDomain

createDomain_domainName :: Lens' CreateDomain Text

The name of the domain to create. The name can range between 3 and 255 characters and can contain the following characters: a-z, A-Z, 0-9, '_', '-', and '.'.

DeleteDomain

deleteDomain_domainName :: Lens' DeleteDomain Text

The name of the domain to delete.

GetAttributes

getAttributes_consistentRead :: Lens' GetAttributes (Maybe Bool)

Determines whether or not strong consistency should be enforced when data is read from SimpleDB. If true, any data previously written to SimpleDB will be returned. Otherwise, results will be consistent eventually, and the client may not see data that was written immediately before your read.

getAttributes_attributeNames :: Lens' GetAttributes (Maybe [Text])

The names of the attributes.

getAttributes_domainName :: Lens' GetAttributes Text

The name of the domain in which to perform the operation.

getAttributes_itemName :: Lens' GetAttributes Text

The name of the item.

getAttributesResponse_attributes :: Lens' GetAttributesResponse (Maybe [Attribute])

The list of attributes returned by the operation.

getAttributesResponse_httpStatus :: Lens' GetAttributesResponse Int

The response's http status code.

PutAttributes

putAttributes_expected :: Lens' PutAttributes (Maybe UpdateCondition)

The update condition which, if specified, determines whether the specified attributes will be updated or not. The update condition must be satisfied in order for this request to be processed and the attributes to be updated.

putAttributes_domainName :: Lens' PutAttributes Text

The name of the domain in which to perform the operation.

putAttributes_itemName :: Lens' PutAttributes Text

The name of the item.

putAttributes_attributes :: Lens' PutAttributes [ReplaceableAttribute]

The list of attributes.

Select

select_nextToken :: Lens' Select (Maybe Text)

A string informing Amazon SimpleDB where to start the next list of ItemNames.

select_consistentRead :: Lens' Select (Maybe Bool)

Determines whether or not strong consistency should be enforced when data is read from SimpleDB. If true, any data previously written to SimpleDB will be returned. Otherwise, results will be consistent eventually, and the client may not see data that was written immediately before your read.

select_selectExpression :: Lens' Select Text

The expression used to query the domain.

selectResponse_nextToken :: Lens' SelectResponse (Maybe Text)

An opaque token indicating that more items than MaxNumberOfItems were matched, the response size exceeded 1 megabyte, or the execution time exceeded 5 seconds.

selectResponse_items :: Lens' SelectResponse (Maybe [Item])

A list of items that match the select expression.

selectResponse_httpStatus :: Lens' SelectResponse Int

The response's http status code.

BatchDeleteAttributes

batchDeleteAttributes_domainName :: Lens' BatchDeleteAttributes Text

The name of the domain in which the attributes are being deleted.

batchDeleteAttributes_items :: Lens' BatchDeleteAttributes [DeletableItem]

A list of items on which to perform the operation.

Types

Attribute

attribute_alternateNameEncoding :: Lens' Attribute (Maybe Text)

attribute_alternateValueEncoding :: Lens' Attribute (Maybe Text)

attribute_name :: Lens' Attribute Text

The name of the attribute.

attribute_value :: Lens' Attribute Text

The value of the attribute.

DeletableItem

deletableItem_attributes :: Lens' DeletableItem (Maybe [Attribute])

Undocumented member.

deletableItem_name :: Lens' DeletableItem Text

Undocumented member.

Item

item_alternateNameEncoding :: Lens' Item (Maybe Text)

item_name :: Lens' Item Text

The name of the item.

item_attributes :: Lens' Item [Attribute]

A list of attributes.

ReplaceableAttribute

replaceableAttribute_replace :: Lens' ReplaceableAttribute (Maybe Bool)

A flag specifying whether or not to replace the attribute/value pair or to add a new attribute/value pair. The default setting is false.

replaceableAttribute_name :: Lens' ReplaceableAttribute Text

The name of the replaceable attribute.

replaceableAttribute_value :: Lens' ReplaceableAttribute Text

The value of the replaceable attribute.

ReplaceableItem

replaceableItem_name :: Lens' ReplaceableItem Text

The name of the replaceable item.

replaceableItem_attributes :: Lens' ReplaceableItem [ReplaceableAttribute]

The list of attributes for a replaceable item.

UpdateCondition

updateCondition_exists :: Lens' UpdateCondition (Maybe Bool)

A value specifying whether or not the specified attribute must exist with the specified value in order for the update condition to be satisfied. Specify true if the attribute must exist for the update condition to be satisfied. Specify false if the attribute should not exist in order for the update condition to be satisfied.

updateCondition_name :: Lens' UpdateCondition (Maybe Text)

The name of the attribute involved in the condition.

updateCondition_value :: Lens' UpdateCondition (Maybe Text)

The value of an attribute. This value can only be specified when the Exists parameter is equal to true.