Network.AWS.SDB.Types

Description

Service Configuration

defaultService :: Service

API version 2009-04-15 of the Amazon SimpleDB SDK configuration.

Errors

_NumberDomainAttributesExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Too many attributes in this domain.

_NumberSubmittedItemsExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Too many items exist in a single call.

_NoSuchDomain :: AsError a => Getting (First ServiceError) a ServiceError

The specified domain does not exist.

_NumberDomainBytesExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Too many bytes in this domain.

_NumberDomainsExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Too many domains exist per this account.

_InvalidNumberPredicates :: AsError a => Getting (First ServiceError) a ServiceError

Too many predicates exist in the query expression.

_DuplicateItemName :: AsError a => Getting (First ServiceError) a ServiceError

The item name was specified more than once.

_AttributeDoesNotExist :: AsError a => Getting (First ServiceError) a ServiceError

The specified attribute does not exist.

_InvalidNumberValueTests :: AsError a => Getting (First ServiceError) a ServiceError

Too many predicates exist in the query expression.

_InvalidQueryExpression :: AsError a => Getting (First ServiceError) a ServiceError

The specified query expression syntax is not valid.

_NumberSubmittedAttributesExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Too many attributes exist in a single call.

_InvalidParameterValue :: AsError a => Getting (First ServiceError) a ServiceError

The value for a parameter is invalid.

_TooManyRequestedAttributes :: AsError a => Getting (First ServiceError) a ServiceError

Too many attributes requested.

_InvalidNextToken :: AsError a => Getting (First ServiceError) a ServiceError

The specified NextToken is not valid.

_MissingParameter :: AsError a => Getting (First ServiceError) a ServiceError

The request must contain the specified missing parameter.

_NumberItemAttributesExceeded :: AsError a => Getting (First ServiceError) a ServiceError

Too many attributes in this item.

_RequestTimeout :: AsError a => Getting (First ServiceError) a ServiceError

A timeout occurred when attempting to query the specified domain with specified query expression.

Attribute

data Attribute

See: newAttribute smart constructor.

Constructors

Attribute'

Fields

alternateNameEncoding :: Maybe Text

alternateValueEncoding :: Maybe Text

name :: Text

The name of the attribute.

value :: Text

The value of the attribute.

newAttribute

Arguments

:: Text
$sel:name:Attribute'

-> Text
$sel:value:Attribute'

-> Attribute

Create a value of Attribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alternateNameEncoding:Attribute', attribute_alternateNameEncoding -

$sel:alternateValueEncoding:Attribute', attribute_alternateValueEncoding -

$sel:name:Attribute', attribute_name - The name of the attribute.

$sel:value:Attribute', attribute_value - The value of the attribute.

attribute_alternateNameEncoding :: Lens' Attribute (Maybe Text)

attribute_alternateValueEncoding :: Lens' Attribute (Maybe Text)

attribute_name :: Lens' Attribute Text

The name of the attribute.

attribute_value :: Lens' Attribute Text

The value of the attribute.

DeletableItem

data DeletableItem

See: newDeletableItem smart constructor.

Constructors

DeletableItem'

Fields

attributes :: Maybe [Attribute]

name :: Text

newDeletableItem

Arguments

:: Text
$sel:name:DeletableItem'

-> DeletableItem

Create a value of DeletableItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:DeletableItem', deletableItem_attributes - Undocumented member.

$sel:name:DeletableItem', deletableItem_name - Undocumented member.

deletableItem_attributes :: Lens' DeletableItem (Maybe [Attribute])

Undocumented member.

deletableItem_name :: Lens' DeletableItem Text

Undocumented member.

Item

data Item

See: newItem smart constructor.

Constructors

Item'

Fields

alternateNameEncoding :: Maybe Text

name :: Text

The name of the item.

attributes :: [Attribute]

A list of attributes.

newItem

Arguments

:: Text
$sel:name:Item'

-> Item

Create a value of Item with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alternateNameEncoding:Item', item_alternateNameEncoding -

$sel:name:Item', item_name - The name of the item.

$sel:attributes:Item', item_attributes - A list of attributes.

item_alternateNameEncoding :: Lens' Item (Maybe Text)

item_name :: Lens' Item Text

The name of the item.

item_attributes :: Lens' Item [Attribute]

A list of attributes.

ReplaceableAttribute

data ReplaceableAttribute

See: newReplaceableAttribute smart constructor.

Constructors

ReplaceableAttribute'

Fields

replace :: Maybe Bool

A flag specifying whether or not to replace the attribute/value pair or to add a new attribute/value pair. The default setting is false.

name :: Text

The name of the replaceable attribute.

value :: Text

The value of the replaceable attribute.

newReplaceableAttribute

Arguments

:: Text
$sel:name:ReplaceableAttribute'

-> Text
$sel:value:ReplaceableAttribute'

-> ReplaceableAttribute

Create a value of ReplaceableAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replace:ReplaceableAttribute', replaceableAttribute_replace - A flag specifying whether or not to replace the attribute/value pair or to add a new attribute/value pair. The default setting is false.

$sel:name:ReplaceableAttribute', replaceableAttribute_name - The name of the replaceable attribute.

$sel:value:ReplaceableAttribute', replaceableAttribute_value - The value of the replaceable attribute.

replaceableAttribute_replace :: Lens' ReplaceableAttribute (Maybe Bool)

A flag specifying whether or not to replace the attribute/value pair or to add a new attribute/value pair. The default setting is false.

replaceableAttribute_name :: Lens' ReplaceableAttribute Text

The name of the replaceable attribute.

replaceableAttribute_value :: Lens' ReplaceableAttribute Text

The value of the replaceable attribute.

ReplaceableItem

data ReplaceableItem

See: newReplaceableItem smart constructor.

Constructors

ReplaceableItem'

Fields

name :: Text

The name of the replaceable item.

attributes :: [ReplaceableAttribute]

The list of attributes for a replaceable item.

newReplaceableItem

Arguments

:: Text
$sel:name:ReplaceableItem'

-> ReplaceableItem

Create a value of ReplaceableItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ReplaceableItem', replaceableItem_name - The name of the replaceable item.

$sel:attributes:ReplaceableItem', replaceableItem_attributes - The list of attributes for a replaceable item.

replaceableItem_name :: Lens' ReplaceableItem Text

The name of the replaceable item.

replaceableItem_attributes :: Lens' ReplaceableItem [ReplaceableAttribute]

The list of attributes for a replaceable item.

UpdateCondition

data UpdateCondition

Specifies the conditions under which data should be updated. If an update condition is specified for a request, the data will only be updated if the condition is satisfied. For example, if an attribute with a specific name and value exists, or if a specific attribute doesn't exist.

See: newUpdateCondition smart constructor.

Constructors

UpdateCondition'

Fields

exists :: Maybe Bool

A value specifying whether or not the specified attribute must exist with the specified value in order for the update condition to be satisfied. Specify true if the attribute must exist for the update condition to be satisfied. Specify false if the attribute should not exist in order for the update condition to be satisfied.

name :: Maybe Text

The name of the attribute involved in the condition.

value :: Maybe Text

The value of an attribute. This value can only be specified when the Exists parameter is equal to true.

newUpdateCondition :: UpdateCondition

Create a value of UpdateCondition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exists:UpdateCondition', updateCondition_exists - A value specifying whether or not the specified attribute must exist with the specified value in order for the update condition to be satisfied. Specify true if the attribute must exist for the update condition to be satisfied. Specify false if the attribute should not exist in order for the update condition to be satisfied.

$sel:name:UpdateCondition', updateCondition_name - The name of the attribute involved in the condition.

$sel:value:UpdateCondition', updateCondition_value - The value of an attribute. This value can only be specified when the Exists parameter is equal to true.

updateCondition_exists :: Lens' UpdateCondition (Maybe Bool)

A value specifying whether or not the specified attribute must exist with the specified value in order for the update condition to be satisfied. Specify true if the attribute must exist for the update condition to be satisfied. Specify false if the attribute should not exist in order for the update condition to be satisfied.

updateCondition_name :: Lens' UpdateCondition (Maybe Text)

The name of the attribute involved in the condition.

updateCondition_value :: Lens' UpdateCondition (Maybe Text)

The value of an attribute. This value can only be specified when the Exists parameter is equal to true.