Network.AWS.SDB.Types.UpdateCondition

Description

Documentation

data UpdateCondition

Specifies the conditions under which data should be updated. If an update condition is specified for a request, the data will only be updated if the condition is satisfied. For example, if an attribute with a specific name and value exists, or if a specific attribute doesn't exist.

See: newUpdateCondition smart constructor.

Constructors

UpdateCondition'

Fields

exists :: Maybe Bool

A value specifying whether or not the specified attribute must exist with the specified value in order for the update condition to be satisfied. Specify true if the attribute must exist for the update condition to be satisfied. Specify false if the attribute should not exist in order for the update condition to be satisfied.

value :: Maybe Text

The value of an attribute. This value can only be specified when the Exists parameter is equal to true.

name :: Maybe Text

The name of the attribute involved in the condition.

newUpdateCondition :: UpdateCondition

Create a value of UpdateCondition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exists:UpdateCondition', updateCondition_exists - A value specifying whether or not the specified attribute must exist with the specified value in order for the update condition to be satisfied. Specify true if the attribute must exist for the update condition to be satisfied. Specify false if the attribute should not exist in order for the update condition to be satisfied.

$sel:value:UpdateCondition', updateCondition_value - The value of an attribute. This value can only be specified when the Exists parameter is equal to true.

$sel:name:UpdateCondition', updateCondition_name - The name of the attribute involved in the condition.

updateCondition_exists :: Lens' UpdateCondition (Maybe Bool)

A value specifying whether or not the specified attribute must exist with the specified value in order for the update condition to be satisfied. Specify true if the attribute must exist for the update condition to be satisfied. Specify false if the attribute should not exist in order for the update condition to be satisfied.

updateCondition_value :: Lens' UpdateCondition (Maybe Text)

The value of an attribute. This value can only be specified when the Exists parameter is equal to true.

updateCondition_name :: Lens' UpdateCondition (Maybe Text)

The name of the attribute involved in the condition.