Network.AWS.SES.CloneReceiptRuleSet

Description

Creates a receipt rule set by cloning an existing one. All receipt rules and configurations are copied to the new receipt rule set and are completely independent of the source rule set.

For information about setting up rule sets, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data CloneReceiptRuleSet

Represents a request to create a receipt rule set by cloning an existing one. You use receipt rule sets to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

See: newCloneReceiptRuleSet smart constructor.

Constructors

CloneReceiptRuleSet'

Fields

ruleSetName :: Text

The name of the rule set to create. The name must:

This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

Start and end with a letter or number.

Contain less than 64 characters.

originalRuleSetName :: Text

The name of the rule set to clone.

newCloneReceiptRuleSet

Arguments

:: Text
$sel:ruleSetName:CloneReceiptRuleSet'

-> Text
$sel:originalRuleSetName:CloneReceiptRuleSet'

-> CloneReceiptRuleSet

Create a value of CloneReceiptRuleSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleSetName:CloneReceiptRuleSet', cloneReceiptRuleSet_ruleSetName - The name of the rule set to create. The name must:

This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

Start and end with a letter or number.

Contain less than 64 characters.

$sel:originalRuleSetName:CloneReceiptRuleSet', cloneReceiptRuleSet_originalRuleSetName - The name of the rule set to clone.

Request Lenses

cloneReceiptRuleSet_ruleSetName :: Lens' CloneReceiptRuleSet Text

The name of the rule set to create. The name must:

This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

Start and end with a letter or number.

Contain less than 64 characters.

cloneReceiptRuleSet_originalRuleSetName :: Lens' CloneReceiptRuleSet Text

The name of the rule set to clone.

Destructuring the Response

data CloneReceiptRuleSetResponse

An empty element returned on a successful request.

See: newCloneReceiptRuleSetResponse smart constructor.

Constructors

CloneReceiptRuleSetResponse'

Fields

httpStatus :: Int

The response's http status code.

newCloneReceiptRuleSetResponse

Arguments

:: Int
$sel:httpStatus:CloneReceiptRuleSetResponse'

-> CloneReceiptRuleSetResponse

Create a value of CloneReceiptRuleSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CloneReceiptRuleSetResponse', cloneReceiptRuleSetResponse_httpStatus - The response's http status code.

Response Lenses

cloneReceiptRuleSetResponse_httpStatus :: Lens' CloneReceiptRuleSetResponse Int

The response's http status code.