Network.AWS.SES.CreateConfigurationSetTrackingOptions

Description

Creates an association between a configuration set and a custom domain for open and click event tracking.

By default, images and links used for tracking open and click events are hosted on domains operated by Amazon SES. You can configure a subdomain of your own to handle these events. For information about using custom domains, see the Amazon SES Developer Guide.

Creating a Request

data CreateConfigurationSetTrackingOptions

Represents a request to create an open and click tracking option object in a configuration set.

See: newCreateConfigurationSetTrackingOptions smart constructor.

Constructors

CreateConfigurationSetTrackingOptions'

Fields

configurationSetName :: Text

The name of the configuration set that the tracking options should be associated with.

trackingOptions :: TrackingOptions

newCreateConfigurationSetTrackingOptions

Arguments

:: Text
$sel:configurationSetName:CreateConfigurationSetTrackingOptions'

-> TrackingOptions
$sel:trackingOptions:CreateConfigurationSetTrackingOptions'

-> CreateConfigurationSetTrackingOptions

Create a value of CreateConfigurationSetTrackingOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationSetName:CreateConfigurationSetTrackingOptions', createConfigurationSetTrackingOptions_configurationSetName - The name of the configuration set that the tracking options should be associated with.

$sel:trackingOptions:CreateConfigurationSetTrackingOptions', createConfigurationSetTrackingOptions_trackingOptions - Undocumented member.

Request Lenses

createConfigurationSetTrackingOptions_configurationSetName :: Lens' CreateConfigurationSetTrackingOptions Text

The name of the configuration set that the tracking options should be associated with.

createConfigurationSetTrackingOptions_trackingOptions :: Lens' CreateConfigurationSetTrackingOptions TrackingOptions

Undocumented member.

Destructuring the Response

data CreateConfigurationSetTrackingOptionsResponse

An empty element returned on a successful request.

See: newCreateConfigurationSetTrackingOptionsResponse smart constructor.

Constructors

CreateConfigurationSetTrackingOptionsResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateConfigurationSetTrackingOptionsResponse

Arguments

:: Int
$sel:httpStatus:CreateConfigurationSetTrackingOptionsResponse'

-> CreateConfigurationSetTrackingOptionsResponse

Create a value of CreateConfigurationSetTrackingOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateConfigurationSetTrackingOptionsResponse', createConfigurationSetTrackingOptionsResponse_httpStatus - The response's http status code.

Response Lenses

createConfigurationSetTrackingOptionsResponse_httpStatus :: Lens' CreateConfigurationSetTrackingOptionsResponse Int

The response's http status code.