Network.AWS.SES.CreateReceiptRule

Description

Creates a receipt rule.

For information about setting up receipt rules, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data CreateReceiptRule

Represents a request to create a receipt rule. You use receipt rules to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

See: newCreateReceiptRule smart constructor.

Constructors

CreateReceiptRule'

Fields

after :: Maybe Text

The name of an existing rule after which the new rule will be placed. If this parameter is null, the new rule will be inserted at the beginning of the rule list.

ruleSetName :: Text

The name of the rule set that the receipt rule will be added to.

rule :: ReceiptRule

A data structure that contains the specified rule's name, actions, recipients, domains, enabled status, scan status, and TLS policy.

newCreateReceiptRule

Arguments

:: Text
$sel:ruleSetName:CreateReceiptRule'

-> ReceiptRule
$sel:rule:CreateReceiptRule'

-> CreateReceiptRule

Create a value of CreateReceiptRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:after:CreateReceiptRule', createReceiptRule_after - The name of an existing rule after which the new rule will be placed. If this parameter is null, the new rule will be inserted at the beginning of the rule list.

$sel:ruleSetName:CreateReceiptRule', createReceiptRule_ruleSetName - The name of the rule set that the receipt rule will be added to.

$sel:rule:CreateReceiptRule', createReceiptRule_rule - A data structure that contains the specified rule's name, actions, recipients, domains, enabled status, scan status, and TLS policy.

Request Lenses

createReceiptRule_after :: Lens' CreateReceiptRule (Maybe Text)

The name of an existing rule after which the new rule will be placed. If this parameter is null, the new rule will be inserted at the beginning of the rule list.

createReceiptRule_ruleSetName :: Lens' CreateReceiptRule Text

The name of the rule set that the receipt rule will be added to.

createReceiptRule_rule :: Lens' CreateReceiptRule ReceiptRule

A data structure that contains the specified rule's name, actions, recipients, domains, enabled status, scan status, and TLS policy.

Destructuring the Response

data CreateReceiptRuleResponse

An empty element returned on a successful request.

See: newCreateReceiptRuleResponse smart constructor.

Constructors

CreateReceiptRuleResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateReceiptRuleResponse

Arguments

:: Int
$sel:httpStatus:CreateReceiptRuleResponse'

-> CreateReceiptRuleResponse

Create a value of CreateReceiptRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateReceiptRuleResponse', createReceiptRuleResponse_httpStatus - The response's http status code.

Response Lenses

createReceiptRuleResponse_httpStatus :: Lens' CreateReceiptRuleResponse Int

The response's http status code.