Network.AWS.SES.CreateTemplate

Description

Creates an email template. Email templates enable you to send personalized email to one or more destinations in a single API operation. For more information, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data CreateTemplate

Represents a request to create an email template. For more information, see the Amazon SES Developer Guide.

See: newCreateTemplate smart constructor.

Constructors

CreateTemplate'

Fields

template :: Template

The content of the email, composed of a subject line, an HTML part, and a text-only part.

newCreateTemplate

Arguments

:: Template
$sel:template:CreateTemplate'

-> CreateTemplate

Create a value of CreateTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:template:CreateTemplate', createTemplate_template - The content of the email, composed of a subject line, an HTML part, and a text-only part.

Request Lenses

createTemplate_template :: Lens' CreateTemplate Template

The content of the email, composed of a subject line, an HTML part, and a text-only part.

Destructuring the Response

data CreateTemplateResponse

See: newCreateTemplateResponse smart constructor.

Constructors

CreateTemplateResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateTemplateResponse

Arguments

:: Int
$sel:httpStatus:CreateTemplateResponse'

-> CreateTemplateResponse

Create a value of CreateTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateTemplateResponse', createTemplateResponse_httpStatus - The response's http status code.

Response Lenses

createTemplateResponse_httpStatus :: Lens' CreateTemplateResponse Int

The response's http status code.