Network.AWS.SES.DeleteIdentityPolicy

Description

Deletes the specified sending authorization policy for the given identity (an email address or a domain). This API returns successfully even if a policy with the specified name does not exist.

This API is for the identity owner only. If you have not verified the identity, this API will return an error.

Sending authorization is a feature that enables an identity owner to authorize other senders to use its identities. For information about using sending authorization, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data DeleteIdentityPolicy

Represents a request to delete a sending authorization policy for an identity. Sending authorization is an Amazon SES feature that enables you to authorize other senders to use your identities. For information, see the Amazon SES Developer Guide.

See: newDeleteIdentityPolicy smart constructor.

Constructors

DeleteIdentityPolicy'

Fields

identity :: Text

The identity that is associated with the policy that you want to delete. You can specify the identity by using its name or by using its Amazon Resource Name (ARN). Examples: user@example.com, example.com, arn:aws:ses:us-east-1:123456789012:identity/example.com.

To successfully call this API, you must own the identity.

policyName :: Text

The name of the policy to be deleted.

newDeleteIdentityPolicy

Arguments

:: Text
$sel:identity:DeleteIdentityPolicy'

-> Text
$sel:policyName:DeleteIdentityPolicy'

-> DeleteIdentityPolicy

Create a value of DeleteIdentityPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identity:DeleteIdentityPolicy', deleteIdentityPolicy_identity - The identity that is associated with the policy that you want to delete. You can specify the identity by using its name or by using its Amazon Resource Name (ARN). Examples: user@example.com, example.com, arn:aws:ses:us-east-1:123456789012:identity/example.com.

To successfully call this API, you must own the identity.

$sel:policyName:DeleteIdentityPolicy', deleteIdentityPolicy_policyName - The name of the policy to be deleted.

Request Lenses

deleteIdentityPolicy_identity :: Lens' DeleteIdentityPolicy Text

The identity that is associated with the policy that you want to delete. You can specify the identity by using its name or by using its Amazon Resource Name (ARN). Examples: user@example.com, example.com, arn:aws:ses:us-east-1:123456789012:identity/example.com.

To successfully call this API, you must own the identity.

deleteIdentityPolicy_policyName :: Lens' DeleteIdentityPolicy Text

The name of the policy to be deleted.

Destructuring the Response

data DeleteIdentityPolicyResponse

An empty element returned on a successful request.

See: newDeleteIdentityPolicyResponse smart constructor.

Constructors

DeleteIdentityPolicyResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteIdentityPolicyResponse

Arguments

:: Int
$sel:httpStatus:DeleteIdentityPolicyResponse'

-> DeleteIdentityPolicyResponse

Create a value of DeleteIdentityPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteIdentityPolicyResponse', deleteIdentityPolicyResponse_httpStatus - The response's http status code.

Response Lenses

deleteIdentityPolicyResponse_httpStatus :: Lens' DeleteIdentityPolicyResponse Int

The response's http status code.