Network.AWS.SES.DescribeActiveReceiptRuleSet

Description

Returns the metadata and receipt rules for the receipt rule set that is currently active.

For information about setting up receipt rule sets, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data DescribeActiveReceiptRuleSet

Represents a request to return the metadata and receipt rules for the receipt rule set that is currently active. You use receipt rule sets to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

See: newDescribeActiveReceiptRuleSet smart constructor.

Constructors

DescribeActiveReceiptRuleSet'

newDescribeActiveReceiptRuleSet :: DescribeActiveReceiptRuleSet

Create a value of DescribeActiveReceiptRuleSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data DescribeActiveReceiptRuleSetResponse

Represents the metadata and receipt rules for the receipt rule set that is currently active.

See: newDescribeActiveReceiptRuleSetResponse smart constructor.

Constructors

DescribeActiveReceiptRuleSetResponse'

Fields

metadata :: Maybe ReceiptRuleSetMetadata

The metadata for the currently active receipt rule set. The metadata consists of the rule set name and a timestamp of when the rule set was created.

rules :: Maybe [ReceiptRule]

The receipt rules that belong to the active rule set.

httpStatus :: Int

The response's http status code.

newDescribeActiveReceiptRuleSetResponse

Arguments

:: Int
$sel:httpStatus:DescribeActiveReceiptRuleSetResponse'

-> DescribeActiveReceiptRuleSetResponse

Create a value of DescribeActiveReceiptRuleSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metadata:DescribeActiveReceiptRuleSetResponse', describeActiveReceiptRuleSetResponse_metadata - The metadata for the currently active receipt rule set. The metadata consists of the rule set name and a timestamp of when the rule set was created.

$sel:rules:DescribeActiveReceiptRuleSetResponse', describeActiveReceiptRuleSetResponse_rules - The receipt rules that belong to the active rule set.

$sel:httpStatus:DescribeActiveReceiptRuleSetResponse', describeActiveReceiptRuleSetResponse_httpStatus - The response's http status code.

Response Lenses

describeActiveReceiptRuleSetResponse_metadata :: Lens' DescribeActiveReceiptRuleSetResponse (Maybe ReceiptRuleSetMetadata)

The metadata for the currently active receipt rule set. The metadata consists of the rule set name and a timestamp of when the rule set was created.

describeActiveReceiptRuleSetResponse_rules :: Lens' DescribeActiveReceiptRuleSetResponse (Maybe [ReceiptRule])

The receipt rules that belong to the active rule set.

describeActiveReceiptRuleSetResponse_httpStatus :: Lens' DescribeActiveReceiptRuleSetResponse Int

The response's http status code.