Network.AWS.SES.DescribeConfigurationSet

Description

Returns the details of the specified configuration set. For information about using configuration sets, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data DescribeConfigurationSet

Represents a request to return the details of a configuration set. Configuration sets enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

See: newDescribeConfigurationSet smart constructor.

Constructors

DescribeConfigurationSet'

Fields

configurationSetAttributeNames :: Maybe [ConfigurationSetAttribute]

A list of configuration set attributes to return.

configurationSetName :: Text

The name of the configuration set to describe.

newDescribeConfigurationSet

Arguments

:: Text
$sel:configurationSetName:DescribeConfigurationSet'

-> DescribeConfigurationSet

Create a value of DescribeConfigurationSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationSetAttributeNames:DescribeConfigurationSet', describeConfigurationSet_configurationSetAttributeNames - A list of configuration set attributes to return.

$sel:configurationSetName:DescribeConfigurationSet', describeConfigurationSet_configurationSetName - The name of the configuration set to describe.

Request Lenses

describeConfigurationSet_configurationSetAttributeNames :: Lens' DescribeConfigurationSet (Maybe [ConfigurationSetAttribute])

A list of configuration set attributes to return.

describeConfigurationSet_configurationSetName :: Lens' DescribeConfigurationSet Text

The name of the configuration set to describe.

Destructuring the Response

data DescribeConfigurationSetResponse

Represents the details of a configuration set. Configuration sets enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

See: newDescribeConfigurationSetResponse smart constructor.

Constructors

DescribeConfigurationSetResponse'

Fields

trackingOptions :: Maybe TrackingOptions

The name of the custom open and click tracking domain associated with the configuration set.

deliveryOptions :: Maybe DeliveryOptions

reputationOptions :: Maybe ReputationOptions

An object that represents the reputation settings for the configuration set.

eventDestinations :: Maybe [EventDestination]

A list of event destinations associated with the configuration set.

configurationSet :: Maybe ConfigurationSet

The configuration set object associated with the specified configuration set.

httpStatus :: Int

The response's http status code.

newDescribeConfigurationSetResponse

Arguments

:: Int
$sel:httpStatus:DescribeConfigurationSetResponse'

-> DescribeConfigurationSetResponse

Create a value of DescribeConfigurationSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trackingOptions:DescribeConfigurationSetResponse', describeConfigurationSetResponse_trackingOptions - The name of the custom open and click tracking domain associated with the configuration set.

$sel:deliveryOptions:DescribeConfigurationSetResponse', describeConfigurationSetResponse_deliveryOptions - Undocumented member.

$sel:reputationOptions:DescribeConfigurationSetResponse', describeConfigurationSetResponse_reputationOptions - An object that represents the reputation settings for the configuration set.

$sel:eventDestinations:DescribeConfigurationSetResponse', describeConfigurationSetResponse_eventDestinations - A list of event destinations associated with the configuration set.

$sel:configurationSet:DescribeConfigurationSetResponse', describeConfigurationSetResponse_configurationSet - The configuration set object associated with the specified configuration set.

$sel:httpStatus:DescribeConfigurationSetResponse', describeConfigurationSetResponse_httpStatus - The response's http status code.

Response Lenses

describeConfigurationSetResponse_trackingOptions :: Lens' DescribeConfigurationSetResponse (Maybe TrackingOptions)

The name of the custom open and click tracking domain associated with the configuration set.

describeConfigurationSetResponse_deliveryOptions :: Lens' DescribeConfigurationSetResponse (Maybe DeliveryOptions)

Undocumented member.

describeConfigurationSetResponse_reputationOptions :: Lens' DescribeConfigurationSetResponse (Maybe ReputationOptions)

An object that represents the reputation settings for the configuration set.

describeConfigurationSetResponse_eventDestinations :: Lens' DescribeConfigurationSetResponse (Maybe [EventDestination])

A list of event destinations associated with the configuration set.

describeConfigurationSetResponse_configurationSet :: Lens' DescribeConfigurationSetResponse (Maybe ConfigurationSet)

The configuration set object associated with the specified configuration set.

describeConfigurationSetResponse_httpStatus :: Lens' DescribeConfigurationSetResponse Int

The response's http status code.