Network.AWS.SES.DescribeReceiptRule

Description

Returns the details of the specified receipt rule.

For information about setting up receipt rules, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data DescribeReceiptRule

Represents a request to return the details of a receipt rule. You use receipt rules to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

See: newDescribeReceiptRule smart constructor.

Constructors

DescribeReceiptRule'

Fields

ruleSetName :: Text

The name of the receipt rule set that the receipt rule belongs to.

ruleName :: Text

The name of the receipt rule.

newDescribeReceiptRule

Arguments

:: Text
$sel:ruleSetName:DescribeReceiptRule'

-> Text
$sel:ruleName:DescribeReceiptRule'

-> DescribeReceiptRule

Create a value of DescribeReceiptRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleSetName:DescribeReceiptRule', describeReceiptRule_ruleSetName - The name of the receipt rule set that the receipt rule belongs to.

$sel:ruleName:DescribeReceiptRule', describeReceiptRule_ruleName - The name of the receipt rule.

Request Lenses

describeReceiptRule_ruleSetName :: Lens' DescribeReceiptRule Text

The name of the receipt rule set that the receipt rule belongs to.

describeReceiptRule_ruleName :: Lens' DescribeReceiptRule Text

The name of the receipt rule.

Destructuring the Response

data DescribeReceiptRuleResponse

Represents the details of a receipt rule.

See: newDescribeReceiptRuleResponse smart constructor.

Constructors

DescribeReceiptRuleResponse'

Fields

rule :: Maybe ReceiptRule

A data structure that contains the specified receipt rule's name, actions, recipients, domains, enabled status, scan status, and Transport Layer Security (TLS) policy.

httpStatus :: Int

The response's http status code.

newDescribeReceiptRuleResponse

Arguments

:: Int
$sel:httpStatus:DescribeReceiptRuleResponse'

-> DescribeReceiptRuleResponse

Create a value of DescribeReceiptRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rule:DescribeReceiptRuleResponse', describeReceiptRuleResponse_rule - A data structure that contains the specified receipt rule's name, actions, recipients, domains, enabled status, scan status, and Transport Layer Security (TLS) policy.

$sel:httpStatus:DescribeReceiptRuleResponse', describeReceiptRuleResponse_httpStatus - The response's http status code.

Response Lenses

describeReceiptRuleResponse_rule :: Lens' DescribeReceiptRuleResponse (Maybe ReceiptRule)

A data structure that contains the specified receipt rule's name, actions, recipients, domains, enabled status, scan status, and Transport Layer Security (TLS) policy.

describeReceiptRuleResponse_httpStatus :: Lens' DescribeReceiptRuleResponse Int

The response's http status code.