Network.AWS.SES.GetSendQuota

Description

Provides the sending limits for the Amazon SES account.

You can execute this operation no more than once per second.

Creating a Request

data GetSendQuota

See: newGetSendQuota smart constructor.

Constructors

GetSendQuota'

newGetSendQuota :: GetSendQuota

Create a value of GetSendQuota with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetSendQuotaResponse

Represents your Amazon SES daily sending quota, maximum send rate, and the number of emails you have sent in the last 24 hours.

See: newGetSendQuotaResponse smart constructor.

Constructors

GetSendQuotaResponse'

Fields

max24HourSend :: Maybe Double

The maximum number of emails the user is allowed to send in a 24-hour interval. A value of -1 signifies an unlimited quota.

sentLast24Hours :: Maybe Double

The number of emails sent during the previous 24 hours.

maxSendRate :: Maybe Double

The maximum number of emails that Amazon SES can accept from the user's account per second.

The rate at which Amazon SES accepts the user's messages might be less than the maximum send rate.

httpStatus :: Int

The response's http status code.

newGetSendQuotaResponse

Arguments

:: Int
$sel:httpStatus:GetSendQuotaResponse'

-> GetSendQuotaResponse

Create a value of GetSendQuotaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:max24HourSend:GetSendQuotaResponse', getSendQuotaResponse_max24HourSend - The maximum number of emails the user is allowed to send in a 24-hour interval. A value of -1 signifies an unlimited quota.

$sel:sentLast24Hours:GetSendQuotaResponse', getSendQuotaResponse_sentLast24Hours - The number of emails sent during the previous 24 hours.

$sel:maxSendRate:GetSendQuotaResponse', getSendQuotaResponse_maxSendRate - The maximum number of emails that Amazon SES can accept from the user's account per second.

The rate at which Amazon SES accepts the user's messages might be less than the maximum send rate.

$sel:httpStatus:GetSendQuotaResponse', getSendQuotaResponse_httpStatus - The response's http status code.

Response Lenses

getSendQuotaResponse_max24HourSend :: Lens' GetSendQuotaResponse (Maybe Double)

The maximum number of emails the user is allowed to send in a 24-hour interval. A value of -1 signifies an unlimited quota.

getSendQuotaResponse_sentLast24Hours :: Lens' GetSendQuotaResponse (Maybe Double)

The number of emails sent during the previous 24 hours.

getSendQuotaResponse_maxSendRate :: Lens' GetSendQuotaResponse (Maybe Double)

The maximum number of emails that Amazon SES can accept from the user's account per second.

The rate at which Amazon SES accepts the user's messages might be less than the maximum send rate.

getSendQuotaResponse_httpStatus :: Lens' GetSendQuotaResponse Int

The response's http status code.