Network.AWS.SES.GetSendStatistics

Description

Provides sending statistics for the current AWS Region. The result is a list of data points, representing the last two weeks of sending activity. Each data point in the list contains statistics for a 15-minute period of time.

You can execute this operation no more than once per second.

Creating a Request

data GetSendStatistics

See: newGetSendStatistics smart constructor.

Constructors

GetSendStatistics'

newGetSendStatistics :: GetSendStatistics

Create a value of GetSendStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetSendStatisticsResponse

Represents a list of data points. This list contains aggregated data from the previous two weeks of your sending activity with Amazon SES.

See: newGetSendStatisticsResponse smart constructor.

Constructors

GetSendStatisticsResponse'

Fields

sendDataPoints :: Maybe [SendDataPoint]

A list of data points, each of which represents 15 minutes of activity.

httpStatus :: Int

The response's http status code.

newGetSendStatisticsResponse

Arguments

:: Int
$sel:httpStatus:GetSendStatisticsResponse'

-> GetSendStatisticsResponse

Create a value of GetSendStatisticsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sendDataPoints:GetSendStatisticsResponse', getSendStatisticsResponse_sendDataPoints - A list of data points, each of which represents 15 minutes of activity.

$sel:httpStatus:GetSendStatisticsResponse', getSendStatisticsResponse_httpStatus - The response's http status code.

Response Lenses

getSendStatisticsResponse_sendDataPoints :: Lens' GetSendStatisticsResponse (Maybe [SendDataPoint])

A list of data points, each of which represents 15 minutes of activity.

getSendStatisticsResponse_httpStatus :: Lens' GetSendStatisticsResponse Int

The response's http status code.