Network.AWS.SES.ListReceiptFilters

Description

Lists the IP address filters associated with your AWS account in the current AWS Region.

For information about managing IP address filters, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data ListReceiptFilters

Represents a request to list the IP address filters that exist under your AWS account. You use IP address filters when you receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

See: newListReceiptFilters smart constructor.

Constructors

ListReceiptFilters'

newListReceiptFilters :: ListReceiptFilters

Create a value of ListReceiptFilters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data ListReceiptFiltersResponse

A list of IP address filters that exist under your AWS account.

See: newListReceiptFiltersResponse smart constructor.

Constructors

ListReceiptFiltersResponse'

Fields

filters :: Maybe [ReceiptFilter]

A list of IP address filter data structures, which each consist of a name, an IP address range, and whether to allow or block mail from it.

httpStatus :: Int

The response's http status code.

newListReceiptFiltersResponse

Arguments

:: Int
$sel:httpStatus:ListReceiptFiltersResponse'

-> ListReceiptFiltersResponse

Create a value of ListReceiptFiltersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListReceiptFiltersResponse', listReceiptFiltersResponse_filters - A list of IP address filter data structures, which each consist of a name, an IP address range, and whether to allow or block mail from it.

$sel:httpStatus:ListReceiptFiltersResponse', listReceiptFiltersResponse_httpStatus - The response's http status code.

Response Lenses

listReceiptFiltersResponse_filters :: Lens' ListReceiptFiltersResponse (Maybe [ReceiptFilter])

A list of IP address filter data structures, which each consist of a name, an IP address range, and whether to allow or block mail from it.

listReceiptFiltersResponse_httpStatus :: Lens' ListReceiptFiltersResponse Int

The response's http status code.