Network.AWS.SES.SendBounce

Description

Generates and sends a bounce message to the sender of an email you received through Amazon SES. You can only use this API on an email up to 24 hours after you receive it.

You cannot use this API to send generic bounces for mail that was not received by Amazon SES.

For information about receiving email through Amazon SES, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data SendBounce

Represents a request to send a bounce message to the sender of an email you received through Amazon SES.

See: newSendBounce smart constructor.

Constructors

SendBounce'

Fields

messageDsn :: Maybe MessageDsn

Message-related DSN fields. If not specified, Amazon SES will choose the values.

explanation :: Maybe Text

Human-readable text for the bounce message to explain the failure. If not specified, the text will be auto-generated based on the bounced recipient information.

bounceSenderArn :: Maybe Text

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the address in the "From" header of the bounce. For more information about sending authorization, see the Amazon SES Developer Guide.

originalMessageId :: Text

The message ID of the message to be bounced.

bounceSender :: Text

The address to use in the "From" header of the bounce message. This must be an identity that you have verified with Amazon SES.

bouncedRecipientInfoList :: [BouncedRecipientInfo]

A list of recipients of the bounced message, including the information required to create the Delivery Status Notifications (DSNs) for the recipients. You must specify at least one BouncedRecipientInfo in the list.

newSendBounce

Arguments

:: Text
$sel:originalMessageId:SendBounce'

-> Text
$sel:bounceSender:SendBounce'

-> SendBounce

Create a value of SendBounce with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:messageDsn:SendBounce', sendBounce_messageDsn - Message-related DSN fields. If not specified, Amazon SES will choose the values.

$sel:explanation:SendBounce', sendBounce_explanation - Human-readable text for the bounce message to explain the failure. If not specified, the text will be auto-generated based on the bounced recipient information.

$sel:bounceSenderArn:SendBounce', sendBounce_bounceSenderArn - This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the address in the "From" header of the bounce. For more information about sending authorization, see the Amazon SES Developer Guide.

$sel:originalMessageId:SendBounce', sendBounce_originalMessageId - The message ID of the message to be bounced.

$sel:bounceSender:SendBounce', sendBounce_bounceSender - The address to use in the "From" header of the bounce message. This must be an identity that you have verified with Amazon SES.

$sel:bouncedRecipientInfoList:SendBounce', sendBounce_bouncedRecipientInfoList - A list of recipients of the bounced message, including the information required to create the Delivery Status Notifications (DSNs) for the recipients. You must specify at least one BouncedRecipientInfo in the list.

Request Lenses

sendBounce_messageDsn :: Lens' SendBounce (Maybe MessageDsn)

Message-related DSN fields. If not specified, Amazon SES will choose the values.

sendBounce_explanation :: Lens' SendBounce (Maybe Text)

Human-readable text for the bounce message to explain the failure. If not specified, the text will be auto-generated based on the bounced recipient information.

sendBounce_bounceSenderArn :: Lens' SendBounce (Maybe Text)

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the address in the "From" header of the bounce. For more information about sending authorization, see the Amazon SES Developer Guide.

sendBounce_originalMessageId :: Lens' SendBounce Text

The message ID of the message to be bounced.

sendBounce_bounceSender :: Lens' SendBounce Text

The address to use in the "From" header of the bounce message. This must be an identity that you have verified with Amazon SES.

sendBounce_bouncedRecipientInfoList :: Lens' SendBounce [BouncedRecipientInfo]

A list of recipients of the bounced message, including the information required to create the Delivery Status Notifications (DSNs) for the recipients. You must specify at least one BouncedRecipientInfo in the list.

Destructuring the Response

data SendBounceResponse

Represents a unique message ID.

See: newSendBounceResponse smart constructor.

Constructors

SendBounceResponse'

Fields

messageId :: Maybe Text

The message ID of the bounce message.

httpStatus :: Int

The response's http status code.

newSendBounceResponse

Arguments

:: Int
$sel:httpStatus:SendBounceResponse'

-> SendBounceResponse

Create a value of SendBounceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:messageId:SendBounceResponse', sendBounceResponse_messageId - The message ID of the bounce message.

$sel:httpStatus:SendBounceResponse', sendBounceResponse_httpStatus - The response's http status code.

Response Lenses

sendBounceResponse_messageId :: Lens' SendBounceResponse (Maybe Text)

The message ID of the bounce message.

sendBounceResponse_httpStatus :: Lens' SendBounceResponse Int

The response's http status code.