Network.AWS.SES.SetIdentityDkimEnabled

Description

Enables or disables Easy DKIM signing of email sent from an identity. If Easy DKIM signing is enabled for a domain, then Amazon SES uses DKIM to sign all email that it sends from addresses on that domain. If Easy DKIM signing is enabled for an email address, then Amazon SES uses DKIM to sign all email it sends from that address.

For email addresses (for example, user@example.com), you can only enable DKIM signing if the corresponding domain (in this case, example.com) has been set up to use Easy DKIM.

You can enable DKIM signing for an identity at any time after you start the verification process for the identity, even if the verification process isn't complete.

You can execute this operation no more than once per second.

For more information about Easy DKIM signing, go to the Amazon SES Developer Guide.

Creating a Request

data SetIdentityDkimEnabled

Represents a request to enable or disable Amazon SES Easy DKIM signing for an identity. For more information about setting up Easy DKIM, see the Amazon SES Developer Guide.

See: newSetIdentityDkimEnabled smart constructor.

Constructors

SetIdentityDkimEnabled'

Fields

identity :: Text

The identity for which DKIM signing should be enabled or disabled.

dkimEnabled :: Bool

Sets whether DKIM signing is enabled for an identity. Set to true to enable DKIM signing for this identity; false to disable it.

newSetIdentityDkimEnabled

Arguments

:: Text
$sel:identity:SetIdentityDkimEnabled'

-> Bool
$sel:dkimEnabled:SetIdentityDkimEnabled'

-> SetIdentityDkimEnabled

Create a value of SetIdentityDkimEnabled with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identity:SetIdentityDkimEnabled', setIdentityDkimEnabled_identity - The identity for which DKIM signing should be enabled or disabled.

$sel:dkimEnabled:SetIdentityDkimEnabled', setIdentityDkimEnabled_dkimEnabled - Sets whether DKIM signing is enabled for an identity. Set to true to enable DKIM signing for this identity; false to disable it.

Request Lenses

setIdentityDkimEnabled_identity :: Lens' SetIdentityDkimEnabled Text

The identity for which DKIM signing should be enabled or disabled.

setIdentityDkimEnabled_dkimEnabled :: Lens' SetIdentityDkimEnabled Bool

Sets whether DKIM signing is enabled for an identity. Set to true to enable DKIM signing for this identity; false to disable it.

Destructuring the Response

data SetIdentityDkimEnabledResponse

An empty element returned on a successful request.

See: newSetIdentityDkimEnabledResponse smart constructor.

Constructors

SetIdentityDkimEnabledResponse'

Fields

httpStatus :: Int

The response's http status code.

newSetIdentityDkimEnabledResponse

Arguments

:: Int
$sel:httpStatus:SetIdentityDkimEnabledResponse'

-> SetIdentityDkimEnabledResponse

Create a value of SetIdentityDkimEnabledResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:SetIdentityDkimEnabledResponse', setIdentityDkimEnabledResponse_httpStatus - The response's http status code.

Response Lenses

setIdentityDkimEnabledResponse_httpStatus :: Lens' SetIdentityDkimEnabledResponse Int

The response's http status code.