Network.AWS.SES.SetIdentityHeadersInNotificationsEnabled

Description

Given an identity (an email address or a domain), sets whether Amazon SES includes the original email headers in the Amazon Simple Notification Service (Amazon SNS) notifications of a specified type.

You can execute this operation no more than once per second.

For more information about using notifications with Amazon SES, see the Amazon SES Developer Guide.

Creating a Request

data SetIdentityHeadersInNotificationsEnabled

Represents a request to set whether Amazon SES includes the original email headers in the Amazon SNS notifications of a specified type. For information about notifications, see the Amazon SES Developer Guide.

See: newSetIdentityHeadersInNotificationsEnabled smart constructor.

Constructors

SetIdentityHeadersInNotificationsEnabled'

Fields

identity :: Text

The identity for which to enable or disable headers in notifications. Examples: user@example.com, example.com.

notificationType :: NotificationType

The notification type for which to enable or disable headers in notifications.

enabled :: Bool

Sets whether Amazon SES includes the original email headers in Amazon SNS notifications of the specified notification type. A value of true specifies that Amazon SES will include headers in notifications, and a value of false specifies that Amazon SES will not include headers in notifications.

This value can only be set when NotificationType is already set to use a particular Amazon SNS topic.

newSetIdentityHeadersInNotificationsEnabled

Arguments

:: Text
$sel:identity:SetIdentityHeadersInNotificationsEnabled'

-> NotificationType
$sel:notificationType:SetIdentityHeadersInNotificationsEnabled'

-> Bool
$sel:enabled:SetIdentityHeadersInNotificationsEnabled'

-> SetIdentityHeadersInNotificationsEnabled

Create a value of SetIdentityHeadersInNotificationsEnabled with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identity:SetIdentityHeadersInNotificationsEnabled', setIdentityHeadersInNotificationsEnabled_identity - The identity for which to enable or disable headers in notifications. Examples: user@example.com, example.com.

$sel:notificationType:SetIdentityHeadersInNotificationsEnabled', setIdentityHeadersInNotificationsEnabled_notificationType - The notification type for which to enable or disable headers in notifications.

$sel:enabled:SetIdentityHeadersInNotificationsEnabled', setIdentityHeadersInNotificationsEnabled_enabled - Sets whether Amazon SES includes the original email headers in Amazon SNS notifications of the specified notification type. A value of true specifies that Amazon SES will include headers in notifications, and a value of false specifies that Amazon SES will not include headers in notifications.

This value can only be set when NotificationType is already set to use a particular Amazon SNS topic.

Request Lenses

setIdentityHeadersInNotificationsEnabled_identity :: Lens' SetIdentityHeadersInNotificationsEnabled Text

The identity for which to enable or disable headers in notifications. Examples: user@example.com, example.com.

setIdentityHeadersInNotificationsEnabled_notificationType :: Lens' SetIdentityHeadersInNotificationsEnabled NotificationType

The notification type for which to enable or disable headers in notifications.

setIdentityHeadersInNotificationsEnabled_enabled :: Lens' SetIdentityHeadersInNotificationsEnabled Bool

Sets whether Amazon SES includes the original email headers in Amazon SNS notifications of the specified notification type. A value of true specifies that Amazon SES will include headers in notifications, and a value of false specifies that Amazon SES will not include headers in notifications.

This value can only be set when NotificationType is already set to use a particular Amazon SNS topic.

Destructuring the Response

data SetIdentityHeadersInNotificationsEnabledResponse

An empty element returned on a successful request.

See: newSetIdentityHeadersInNotificationsEnabledResponse smart constructor.

Constructors

SetIdentityHeadersInNotificationsEnabledResponse'

Fields

httpStatus :: Int

The response's http status code.

newSetIdentityHeadersInNotificationsEnabledResponse

Arguments

:: Int
$sel:httpStatus:SetIdentityHeadersInNotificationsEnabledResponse'

-> SetIdentityHeadersInNotificationsEnabledResponse

Create a value of SetIdentityHeadersInNotificationsEnabledResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:SetIdentityHeadersInNotificationsEnabledResponse', setIdentityHeadersInNotificationsEnabledResponse_httpStatus - The response's http status code.

Response Lenses

setIdentityHeadersInNotificationsEnabledResponse_httpStatus :: Lens' SetIdentityHeadersInNotificationsEnabledResponse Int

The response's http status code.