Network.AWS.SES.TestRenderTemplate

Description

Creates a preview of the MIME content of an email when provided with a template and a set of replacement data.

You can execute this operation no more than once per second.

Creating a Request

data TestRenderTemplate

See: newTestRenderTemplate smart constructor.

Constructors

TestRenderTemplate'

Fields

templateName :: Text

The name of the template that you want to render.

templateData :: Text

A list of replacement values to apply to the template. This parameter is a JSON object, typically consisting of key-value pairs in which the keys correspond to replacement tags in the email template.

newTestRenderTemplate

Arguments

:: Text
$sel:templateName:TestRenderTemplate'

-> Text
$sel:templateData:TestRenderTemplate'

-> TestRenderTemplate

Create a value of TestRenderTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:TestRenderTemplate', testRenderTemplate_templateName - The name of the template that you want to render.

$sel:templateData:TestRenderTemplate', testRenderTemplate_templateData - A list of replacement values to apply to the template. This parameter is a JSON object, typically consisting of key-value pairs in which the keys correspond to replacement tags in the email template.

Request Lenses

testRenderTemplate_templateName :: Lens' TestRenderTemplate Text

The name of the template that you want to render.

testRenderTemplate_templateData :: Lens' TestRenderTemplate Text

A list of replacement values to apply to the template. This parameter is a JSON object, typically consisting of key-value pairs in which the keys correspond to replacement tags in the email template.

Destructuring the Response

data TestRenderTemplateResponse

See: newTestRenderTemplateResponse smart constructor.

Constructors

TestRenderTemplateResponse'

Fields

renderedTemplate :: Maybe Text

The complete MIME message rendered by applying the data in the TemplateData parameter to the template specified in the TemplateName parameter.

httpStatus :: Int

The response's http status code.

newTestRenderTemplateResponse

Arguments

:: Int
$sel:httpStatus:TestRenderTemplateResponse'

-> TestRenderTemplateResponse

Create a value of TestRenderTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:renderedTemplate:TestRenderTemplateResponse', testRenderTemplateResponse_renderedTemplate - The complete MIME message rendered by applying the data in the TemplateData parameter to the template specified in the TemplateName parameter.

$sel:httpStatus:TestRenderTemplateResponse', testRenderTemplateResponse_httpStatus - The response's http status code.

Response Lenses

testRenderTemplateResponse_renderedTemplate :: Lens' TestRenderTemplateResponse (Maybe Text)

The complete MIME message rendered by applying the data in the TemplateData parameter to the template specified in the TemplateName parameter.

testRenderTemplateResponse_httpStatus :: Lens' TestRenderTemplateResponse Int

The response's http status code.