Network.AWS.SES.Types.ReceiptAction

Description

Documentation

data ReceiptAction

An action that Amazon SES can take when it receives an email on behalf of one or more email addresses or domains that you own. An instance of this data type can represent only one action.

For information about setting up receipt rules, see the Amazon SES Developer Guide.

See: newReceiptAction smart constructor.

Constructors

ReceiptAction'

Fields

lambdaAction :: Maybe LambdaAction

Calls an AWS Lambda function, and optionally, publishes a notification to Amazon SNS.

s3Action :: Maybe S3Action

Saves the received message to an Amazon Simple Storage Service (Amazon S3) bucket and, optionally, publishes a notification to Amazon SNS.

stopAction :: Maybe StopAction

Terminates the evaluation of the receipt rule set and optionally publishes a notification to Amazon SNS.

workmailAction :: Maybe WorkmailAction

Calls Amazon WorkMail and, optionally, publishes a notification to Amazon Amazon SNS.

bounceAction :: Maybe BounceAction

Rejects the received email by returning a bounce response to the sender and, optionally, publishes a notification to Amazon Simple Notification Service (Amazon SNS).

addHeaderAction :: Maybe AddHeaderAction

Adds a header to the received email.

sNSAction :: Maybe SNSAction

Publishes the email content within a notification to Amazon SNS.

newReceiptAction :: ReceiptAction

Create a value of ReceiptAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lambdaAction:ReceiptAction', receiptAction_lambdaAction - Calls an AWS Lambda function, and optionally, publishes a notification to Amazon SNS.

$sel:s3Action:ReceiptAction', receiptAction_s3Action - Saves the received message to an Amazon Simple Storage Service (Amazon S3) bucket and, optionally, publishes a notification to Amazon SNS.

$sel:stopAction:ReceiptAction', receiptAction_stopAction - Terminates the evaluation of the receipt rule set and optionally publishes a notification to Amazon SNS.

$sel:workmailAction:ReceiptAction', receiptAction_workmailAction - Calls Amazon WorkMail and, optionally, publishes a notification to Amazon Amazon SNS.

$sel:bounceAction:ReceiptAction', receiptAction_bounceAction - Rejects the received email by returning a bounce response to the sender and, optionally, publishes a notification to Amazon Simple Notification Service (Amazon SNS).

$sel:addHeaderAction:ReceiptAction', receiptAction_addHeaderAction - Adds a header to the received email.

$sel:sNSAction:ReceiptAction', receiptAction_sNSAction - Publishes the email content within a notification to Amazon SNS.

receiptAction_lambdaAction :: Lens' ReceiptAction (Maybe LambdaAction)

Calls an AWS Lambda function, and optionally, publishes a notification to Amazon SNS.

receiptAction_s3Action :: Lens' ReceiptAction (Maybe S3Action)

Saves the received message to an Amazon Simple Storage Service (Amazon S3) bucket and, optionally, publishes a notification to Amazon SNS.

receiptAction_stopAction :: Lens' ReceiptAction (Maybe StopAction)

Terminates the evaluation of the receipt rule set and optionally publishes a notification to Amazon SNS.

receiptAction_workmailAction :: Lens' ReceiptAction (Maybe WorkmailAction)

Calls Amazon WorkMail and, optionally, publishes a notification to Amazon Amazon SNS.

receiptAction_bounceAction :: Lens' ReceiptAction (Maybe BounceAction)

Rejects the received email by returning a bounce response to the sender and, optionally, publishes a notification to Amazon Simple Notification Service (Amazon SNS).

receiptAction_addHeaderAction :: Lens' ReceiptAction (Maybe AddHeaderAction)

Adds a header to the received email.

receiptAction_sNSAction :: Lens' ReceiptAction (Maybe SNSAction)

Publishes the email content within a notification to Amazon SNS.