Network.AWS.SES.Types.ReceiptFilter

Description

Documentation

data ReceiptFilter

A receipt IP address filter enables you to specify whether to accept or reject mail originating from an IP address or range of IP addresses.

For information about setting up IP address filters, see the Amazon SES Developer Guide.

See: newReceiptFilter smart constructor.

Constructors

ReceiptFilter'

Fields

name :: Text

The name of the IP address filter. The name must:

This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

Start and end with a letter or number.

Contain less than 64 characters.

ipFilter :: ReceiptIpFilter

A structure that provides the IP addresses to block or allow, and whether to block or allow incoming mail from them.

newReceiptFilter

Arguments

:: Text
$sel:name:ReceiptFilter'

-> ReceiptIpFilter
$sel:ipFilter:ReceiptFilter'

-> ReceiptFilter

Create a value of ReceiptFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ReceiptFilter', receiptFilter_name - The name of the IP address filter. The name must:

This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

Start and end with a letter or number.

Contain less than 64 characters.

$sel:ipFilter:ReceiptFilter', receiptFilter_ipFilter - A structure that provides the IP addresses to block or allow, and whether to block or allow incoming mail from them.

receiptFilter_name :: Lens' ReceiptFilter Text

The name of the IP address filter. The name must:

This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

Start and end with a letter or number.

Contain less than 64 characters.

receiptFilter_ipFilter :: Lens' ReceiptFilter ReceiptIpFilter

A structure that provides the IP addresses to block or allow, and whether to block or allow incoming mail from them.