Network.AWS.SES.Types.ReceiptIpFilter

Description

Documentation

data ReceiptIpFilter

A receipt IP address filter enables you to specify whether to accept or reject mail originating from an IP address or range of IP addresses.

For information about setting up IP address filters, see the Amazon SES Developer Guide.

See: newReceiptIpFilter smart constructor.

Constructors

ReceiptIpFilter'

Fields

policy :: ReceiptFilterPolicy

Indicates whether to block or allow incoming mail from the specified IP addresses.

cidr :: Text

A single IP address or a range of IP addresses that you want to block or allow, specified in Classless Inter-Domain Routing (CIDR) notation. An example of a single email address is 10.0.0.1. An example of a range of IP addresses is 10.0.0.1/24. For more information about CIDR notation, see RFC 2317.

newReceiptIpFilter

Arguments

:: ReceiptFilterPolicy
$sel:policy:ReceiptIpFilter'

-> Text
$sel:cidr:ReceiptIpFilter'

-> ReceiptIpFilter

Create a value of ReceiptIpFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policy:ReceiptIpFilter', receiptIpFilter_policy - Indicates whether to block or allow incoming mail from the specified IP addresses.

$sel:cidr:ReceiptIpFilter', receiptIpFilter_cidr - A single IP address or a range of IP addresses that you want to block or allow, specified in Classless Inter-Domain Routing (CIDR) notation. An example of a single email address is 10.0.0.1. An example of a range of IP addresses is 10.0.0.1/24. For more information about CIDR notation, see RFC 2317.

receiptIpFilter_policy :: Lens' ReceiptIpFilter ReceiptFilterPolicy

Indicates whether to block or allow incoming mail from the specified IP addresses.

receiptIpFilter_cidr :: Lens' ReceiptIpFilter Text

A single IP address or a range of IP addresses that you want to block or allow, specified in Classless Inter-Domain Routing (CIDR) notation. An example of a single email address is 10.0.0.1. An example of a range of IP addresses is 10.0.0.1/24. For more information about CIDR notation, see RFC 2317.