Network.AWS.SES.Types.ReceiptRule

Description

Documentation

data ReceiptRule

Receipt rules enable you to specify which actions Amazon SES should take when it receives mail on behalf of one or more email addresses or domains that you own.

Each receipt rule defines a set of email addresses or domains that it applies to. If the email addresses or domains match at least one recipient address of the message, Amazon SES executes all of the receipt rule's actions on the message.

For information about setting up receipt rules, see the Amazon SES Developer Guide.

See: newReceiptRule smart constructor.

Constructors

ReceiptRule'

Fields

tlsPolicy :: Maybe TlsPolicy

Specifies whether Amazon SES should require that incoming email is delivered over a connection encrypted with Transport Layer Security (TLS). If this parameter is set to Require, Amazon SES will bounce emails that are not received over TLS. The default is Optional.

enabled :: Maybe Bool

If true, the receipt rule is active. The default value is false.

actions :: Maybe [ReceiptAction]

An ordered list of actions to perform on messages that match at least one of the recipient email addresses or domains specified in the receipt rule.

recipients :: Maybe [Text]

The recipient domains and email addresses that the receipt rule applies to. If this field is not specified, this rule will match all recipients under all verified domains.

scanEnabled :: Maybe Bool

If true, then messages that this receipt rule applies to are scanned for spam and viruses. The default value is false.

name :: Text

The name of the receipt rule. The name must:

This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

Start and end with a letter or number.

Contain less than 64 characters.

newReceiptRule

Arguments

:: Text
$sel:name:ReceiptRule'

-> ReceiptRule

Create a value of ReceiptRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tlsPolicy:ReceiptRule', receiptRule_tlsPolicy - Specifies whether Amazon SES should require that incoming email is delivered over a connection encrypted with Transport Layer Security (TLS). If this parameter is set to Require, Amazon SES will bounce emails that are not received over TLS. The default is Optional.

$sel:enabled:ReceiptRule', receiptRule_enabled - If true, the receipt rule is active. The default value is false.

$sel:actions:ReceiptRule', receiptRule_actions - An ordered list of actions to perform on messages that match at least one of the recipient email addresses or domains specified in the receipt rule.

$sel:recipients:ReceiptRule', receiptRule_recipients - The recipient domains and email addresses that the receipt rule applies to. If this field is not specified, this rule will match all recipients under all verified domains.

$sel:scanEnabled:ReceiptRule', receiptRule_scanEnabled - If true, then messages that this receipt rule applies to are scanned for spam and viruses. The default value is false.

$sel:name:ReceiptRule', receiptRule_name - The name of the receipt rule. The name must:

This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

Start and end with a letter or number.

Contain less than 64 characters.

receiptRule_tlsPolicy :: Lens' ReceiptRule (Maybe TlsPolicy)

Specifies whether Amazon SES should require that incoming email is delivered over a connection encrypted with Transport Layer Security (TLS). If this parameter is set to Require, Amazon SES will bounce emails that are not received over TLS. The default is Optional.

receiptRule_enabled :: Lens' ReceiptRule (Maybe Bool)

If true, the receipt rule is active. The default value is false.

receiptRule_actions :: Lens' ReceiptRule (Maybe [ReceiptAction])

An ordered list of actions to perform on messages that match at least one of the recipient email addresses or domains specified in the receipt rule.

receiptRule_recipients :: Lens' ReceiptRule (Maybe [Text])

The recipient domains and email addresses that the receipt rule applies to. If this field is not specified, this rule will match all recipients under all verified domains.

receiptRule_scanEnabled :: Lens' ReceiptRule (Maybe Bool)

If true, then messages that this receipt rule applies to are scanned for spam and viruses. The default value is false.

receiptRule_name :: Lens' ReceiptRule Text

The name of the receipt rule. The name must:

This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

Start and end with a letter or number.

Contain less than 64 characters.