Network.AWS.SES.Types.RecipientDsnFields

Description

Documentation

data RecipientDsnFields

Recipient-related information to include in the Delivery Status Notification (DSN) when an email that Amazon SES receives on your behalf bounces.

For information about receiving email through Amazon SES, see the Amazon SES Developer Guide.

See: newRecipientDsnFields smart constructor.

Constructors

RecipientDsnFields'

Fields

remoteMta :: Maybe Text

The MTA to which the remote MTA attempted to deliver the message, formatted as specified in RFC 3464 (mta-name-type; mta-name). This parameter typically applies only to propagating synchronous bounces.

lastAttemptDate :: Maybe ISO8601

The time the final delivery attempt was made, in RFC 822 date-time format.

extensionFields :: Maybe [ExtensionField]

Additional X-headers to include in the DSN.

diagnosticCode :: Maybe Text

An extended explanation of what went wrong; this is usually an SMTP response. See RFC 3463 for the correct formatting of this parameter.

finalRecipient :: Maybe Text

The email address that the message was ultimately delivered to. This corresponds to the Final-Recipient in the DSN. If not specified, FinalRecipient will be set to the Recipient specified in the BouncedRecipientInfo structure. Either FinalRecipient or the recipient in BouncedRecipientInfo must be a recipient of the original bounced message.

Do not prepend the FinalRecipient email address with rfc 822;, as described in RFC 3798.

action :: DsnAction

The action performed by the reporting mail transfer agent (MTA) as a result of its attempt to deliver the message to the recipient address. This is required by RFC 3464.

status :: Text

The status code that indicates what went wrong. This is required by RFC 3464.

newRecipientDsnFields

Arguments

:: DsnAction
$sel:action:RecipientDsnFields'

-> Text
$sel:status:RecipientDsnFields'

-> RecipientDsnFields

Create a value of RecipientDsnFields with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remoteMta:RecipientDsnFields', recipientDsnFields_remoteMta - The MTA to which the remote MTA attempted to deliver the message, formatted as specified in RFC 3464 (mta-name-type; mta-name). This parameter typically applies only to propagating synchronous bounces.

$sel:lastAttemptDate:RecipientDsnFields', recipientDsnFields_lastAttemptDate - The time the final delivery attempt was made, in RFC 822 date-time format.

$sel:extensionFields:RecipientDsnFields', recipientDsnFields_extensionFields - Additional X-headers to include in the DSN.

$sel:diagnosticCode:RecipientDsnFields', recipientDsnFields_diagnosticCode - An extended explanation of what went wrong; this is usually an SMTP response. See RFC 3463 for the correct formatting of this parameter.

$sel:finalRecipient:RecipientDsnFields', recipientDsnFields_finalRecipient - The email address that the message was ultimately delivered to. This corresponds to the Final-Recipient in the DSN. If not specified, FinalRecipient will be set to the Recipient specified in the BouncedRecipientInfo structure. Either FinalRecipient or the recipient in BouncedRecipientInfo must be a recipient of the original bounced message.

Do not prepend the FinalRecipient email address with rfc 822;, as described in RFC 3798.

$sel:action:RecipientDsnFields', recipientDsnFields_action - The action performed by the reporting mail transfer agent (MTA) as a result of its attempt to deliver the message to the recipient address. This is required by RFC 3464.

$sel:status:RecipientDsnFields', recipientDsnFields_status - The status code that indicates what went wrong. This is required by RFC 3464.

recipientDsnFields_remoteMta :: Lens' RecipientDsnFields (Maybe Text)

The MTA to which the remote MTA attempted to deliver the message, formatted as specified in RFC 3464 (mta-name-type; mta-name). This parameter typically applies only to propagating synchronous bounces.

recipientDsnFields_lastAttemptDate :: Lens' RecipientDsnFields (Maybe UTCTime)

The time the final delivery attempt was made, in RFC 822 date-time format.

recipientDsnFields_extensionFields :: Lens' RecipientDsnFields (Maybe [ExtensionField])

Additional X-headers to include in the DSN.

recipientDsnFields_diagnosticCode :: Lens' RecipientDsnFields (Maybe Text)

An extended explanation of what went wrong; this is usually an SMTP response. See RFC 3463 for the correct formatting of this parameter.

recipientDsnFields_finalRecipient :: Lens' RecipientDsnFields (Maybe Text)

The email address that the message was ultimately delivered to. This corresponds to the Final-Recipient in the DSN. If not specified, FinalRecipient will be set to the Recipient specified in the BouncedRecipientInfo structure. Either FinalRecipient or the recipient in BouncedRecipientInfo must be a recipient of the original bounced message.

Do not prepend the FinalRecipient email address with rfc 822;, as described in RFC 3798.

recipientDsnFields_action :: Lens' RecipientDsnFields DsnAction

The action performed by the reporting mail transfer agent (MTA) as a result of its attempt to deliver the message to the recipient address. This is required by RFC 3464.

recipientDsnFields_status :: Lens' RecipientDsnFields Text

The status code that indicates what went wrong. This is required by RFC 3464.