Network.AWS.SES.Types.ReputationOptions

Description

Documentation

data ReputationOptions

Contains information about the reputation settings for a configuration set.

See: newReputationOptions smart constructor.

Constructors

ReputationOptions'

Fields

reputationMetricsEnabled :: Maybe Bool

Describes whether or not Amazon SES publishes reputation metrics for the configuration set, such as bounce and complaint rates, to Amazon CloudWatch.

If the value is true, reputation metrics are published. If the value is false, reputation metrics are not published. The default value is false.

lastFreshStart :: Maybe ISO8601

The date and time at which the reputation metrics for the configuration set were last reset. Resetting these metrics is known as a /fresh start/.

When you disable email sending for a configuration set using UpdateConfigurationSetSendingEnabled and later re-enable it, the reputation metrics for the configuration set (but not for the entire Amazon SES account) are reset.

If email sending for the configuration set has never been disabled and later re-enabled, the value of this attribute is null.

sendingEnabled :: Maybe Bool

Describes whether email sending is enabled or disabled for the configuration set. If the value is true, then Amazon SES will send emails that use the configuration set. If the value is false, Amazon SES will not send emails that use the configuration set. The default value is true. You can change this setting using UpdateConfigurationSetSendingEnabled.

newReputationOptions :: ReputationOptions

Create a value of ReputationOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reputationMetricsEnabled:ReputationOptions', reputationOptions_reputationMetricsEnabled - Describes whether or not Amazon SES publishes reputation metrics for the configuration set, such as bounce and complaint rates, to Amazon CloudWatch.

If the value is true, reputation metrics are published. If the value is false, reputation metrics are not published. The default value is false.

$sel:lastFreshStart:ReputationOptions', reputationOptions_lastFreshStart - The date and time at which the reputation metrics for the configuration set were last reset. Resetting these metrics is known as a /fresh start/.

When you disable email sending for a configuration set using UpdateConfigurationSetSendingEnabled and later re-enable it, the reputation metrics for the configuration set (but not for the entire Amazon SES account) are reset.

If email sending for the configuration set has never been disabled and later re-enabled, the value of this attribute is null.

$sel:sendingEnabled:ReputationOptions', reputationOptions_sendingEnabled - Describes whether email sending is enabled or disabled for the configuration set. If the value is true, then Amazon SES will send emails that use the configuration set. If the value is false, Amazon SES will not send emails that use the configuration set. The default value is true. You can change this setting using UpdateConfigurationSetSendingEnabled.

reputationOptions_reputationMetricsEnabled :: Lens' ReputationOptions (Maybe Bool)

Describes whether or not Amazon SES publishes reputation metrics for the configuration set, such as bounce and complaint rates, to Amazon CloudWatch.

If the value is true, reputation metrics are published. If the value is false, reputation metrics are not published. The default value is false.

reputationOptions_lastFreshStart :: Lens' ReputationOptions (Maybe UTCTime)

The date and time at which the reputation metrics for the configuration set were last reset. Resetting these metrics is known as a /fresh start/.

When you disable email sending for a configuration set using UpdateConfigurationSetSendingEnabled and later re-enable it, the reputation metrics for the configuration set (but not for the entire Amazon SES account) are reset.

If email sending for the configuration set has never been disabled and later re-enabled, the value of this attribute is null.

reputationOptions_sendingEnabled :: Lens' ReputationOptions (Maybe Bool)

Describes whether email sending is enabled or disabled for the configuration set. If the value is true, then Amazon SES will send emails that use the configuration set. If the value is false, Amazon SES will not send emails that use the configuration set. The default value is true. You can change this setting using UpdateConfigurationSetSendingEnabled.