Network.AWS.SES.Types.SendDataPoint

Description

Documentation

data SendDataPoint

Represents sending statistics data. Each SendDataPoint contains statistics for a 15-minute period of sending activity.

See: newSendDataPoint smart constructor.

Constructors

SendDataPoint'

Fields

bounces :: Maybe Integer

Number of emails that have bounced.

complaints :: Maybe Integer

Number of unwanted emails that were rejected by recipients.

rejects :: Maybe Integer

Number of emails rejected by Amazon SES.

timestamp :: Maybe ISO8601

Time of the data point.

deliveryAttempts :: Maybe Integer

Number of emails that have been sent.

newSendDataPoint :: SendDataPoint

Create a value of SendDataPoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bounces:SendDataPoint', sendDataPoint_bounces - Number of emails that have bounced.

$sel:complaints:SendDataPoint', sendDataPoint_complaints - Number of unwanted emails that were rejected by recipients.

$sel:rejects:SendDataPoint', sendDataPoint_rejects - Number of emails rejected by Amazon SES.

$sel:timestamp:SendDataPoint', sendDataPoint_timestamp - Time of the data point.

$sel:deliveryAttempts:SendDataPoint', sendDataPoint_deliveryAttempts - Number of emails that have been sent.

sendDataPoint_bounces :: Lens' SendDataPoint (Maybe Integer)

Number of emails that have bounced.

sendDataPoint_complaints :: Lens' SendDataPoint (Maybe Integer)

Number of unwanted emails that were rejected by recipients.

sendDataPoint_rejects :: Lens' SendDataPoint (Maybe Integer)

Number of emails rejected by Amazon SES.

sendDataPoint_timestamp :: Lens' SendDataPoint (Maybe UTCTime)

Time of the data point.

sendDataPoint_deliveryAttempts :: Lens' SendDataPoint (Maybe Integer)

Number of emails that have been sent.