Network.AWS.SES.Types.Template

Description

Documentation

data Template

The content of the email, composed of a subject line, an HTML part, and a text-only part.

See: newTemplate smart constructor.

Constructors

Template'

Fields

textPart :: Maybe Text

The email body that will be visible to recipients whose email clients do not display HTML.

subjectPart :: Maybe Text

The subject line of the email.

htmlPart :: Maybe Text

The HTML body of the email.

templateName :: Text

The name of the template. You will refer to this name when you send email using the SendTemplatedEmail or SendBulkTemplatedEmail operations.

newTemplate

Arguments

:: Text
$sel:templateName:Template'

-> Template

Create a value of Template with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:textPart:Template', template_textPart - The email body that will be visible to recipients whose email clients do not display HTML.

$sel:subjectPart:Template', template_subjectPart - The subject line of the email.

$sel:htmlPart:Template', template_htmlPart - The HTML body of the email.

$sel:templateName:Template', template_templateName - The name of the template. You will refer to this name when you send email using the SendTemplatedEmail or SendBulkTemplatedEmail operations.

template_textPart :: Lens' Template (Maybe Text)

The email body that will be visible to recipients whose email clients do not display HTML.

template_subjectPart :: Lens' Template (Maybe Text)

The subject line of the email.

template_htmlPart :: Lens' Template (Maybe Text)

The HTML body of the email.

template_templateName :: Lens' Template Text

The name of the template. You will refer to this name when you send email using the SendTemplatedEmail or SendBulkTemplatedEmail operations.