Network.AWS.SES.UpdateAccountSendingEnabled

Description

Enables or disables email sending across your entire Amazon SES account in the current AWS Region. You can use this operation in conjunction with Amazon CloudWatch alarms to temporarily pause email sending across your Amazon SES account in a given AWS Region when reputation metrics (such as your bounce or complaint rates) reach certain thresholds.

You can execute this operation no more than once per second.

Creating a Request

data UpdateAccountSendingEnabled

Represents a request to enable or disable the email sending capabilities for your entire Amazon SES account.

See: newUpdateAccountSendingEnabled smart constructor.

Constructors

UpdateAccountSendingEnabled'

Fields

enabled :: Maybe Bool

Describes whether email sending is enabled or disabled for your Amazon SES account in the current AWS Region.

newUpdateAccountSendingEnabled :: UpdateAccountSendingEnabled

Create a value of UpdateAccountSendingEnabled with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:UpdateAccountSendingEnabled', updateAccountSendingEnabled_enabled - Describes whether email sending is enabled or disabled for your Amazon SES account in the current AWS Region.

Request Lenses

updateAccountSendingEnabled_enabled :: Lens' UpdateAccountSendingEnabled (Maybe Bool)

Describes whether email sending is enabled or disabled for your Amazon SES account in the current AWS Region.

Destructuring the Response

data UpdateAccountSendingEnabledResponse

See: newUpdateAccountSendingEnabledResponse smart constructor.

Constructors

UpdateAccountSendingEnabledResponse'

newUpdateAccountSendingEnabledResponse :: UpdateAccountSendingEnabledResponse

Create a value of UpdateAccountSendingEnabledResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.