Network.AWS.SES.UpdateConfigurationSetReputationMetricsEnabled

Description

Enables or disables the publishing of reputation metrics for emails sent using a specific configuration set in a given AWS Region. Reputation metrics include bounce and complaint rates. These metrics are published to Amazon CloudWatch. By using CloudWatch, you can create alarms when bounce or complaint rates exceed certain thresholds.

You can execute this operation no more than once per second.

Creating a Request

data UpdateConfigurationSetReputationMetricsEnabled

Represents a request to modify the reputation metric publishing settings for a configuration set.

See: newUpdateConfigurationSetReputationMetricsEnabled smart constructor.

Constructors

UpdateConfigurationSetReputationMetricsEnabled'

Fields

configurationSetName :: Text

The name of the configuration set that you want to update.

enabled :: Bool

Describes whether or not Amazon SES will publish reputation metrics for the configuration set, such as bounce and complaint rates, to Amazon CloudWatch.

newUpdateConfigurationSetReputationMetricsEnabled

Arguments

:: Text
$sel:configurationSetName:UpdateConfigurationSetReputationMetricsEnabled'

-> Bool
$sel:enabled:UpdateConfigurationSetReputationMetricsEnabled'

-> UpdateConfigurationSetReputationMetricsEnabled

Create a value of UpdateConfigurationSetReputationMetricsEnabled with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationSetName:UpdateConfigurationSetReputationMetricsEnabled', updateConfigurationSetReputationMetricsEnabled_configurationSetName - The name of the configuration set that you want to update.

$sel:enabled:UpdateConfigurationSetReputationMetricsEnabled', updateConfigurationSetReputationMetricsEnabled_enabled - Describes whether or not Amazon SES will publish reputation metrics for the configuration set, such as bounce and complaint rates, to Amazon CloudWatch.

Request Lenses

updateConfigurationSetReputationMetricsEnabled_configurationSetName :: Lens' UpdateConfigurationSetReputationMetricsEnabled Text

The name of the configuration set that you want to update.

updateConfigurationSetReputationMetricsEnabled_enabled :: Lens' UpdateConfigurationSetReputationMetricsEnabled Bool

Describes whether or not Amazon SES will publish reputation metrics for the configuration set, such as bounce and complaint rates, to Amazon CloudWatch.

Destructuring the Response

data UpdateConfigurationSetReputationMetricsEnabledResponse

See: newUpdateConfigurationSetReputationMetricsEnabledResponse smart constructor.

Constructors

UpdateConfigurationSetReputationMetricsEnabledResponse'

newUpdateConfigurationSetReputationMetricsEnabledResponse :: UpdateConfigurationSetReputationMetricsEnabledResponse

Create a value of UpdateConfigurationSetReputationMetricsEnabledResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.