Network.AWS.SES.UpdateReceiptRule

Description

Updates a receipt rule.

For information about managing receipt rules, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data UpdateReceiptRule

Represents a request to update a receipt rule. You use receipt rules to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

See: newUpdateReceiptRule smart constructor.

Constructors

UpdateReceiptRule'

Fields

ruleSetName :: Text

The name of the receipt rule set that the receipt rule belongs to.

rule :: ReceiptRule

A data structure that contains the updated receipt rule information.

newUpdateReceiptRule

Arguments

:: Text
$sel:ruleSetName:UpdateReceiptRule'

-> ReceiptRule
$sel:rule:UpdateReceiptRule'

-> UpdateReceiptRule

Create a value of UpdateReceiptRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleSetName:UpdateReceiptRule', updateReceiptRule_ruleSetName - The name of the receipt rule set that the receipt rule belongs to.

$sel:rule:UpdateReceiptRule', updateReceiptRule_rule - A data structure that contains the updated receipt rule information.

Request Lenses

updateReceiptRule_ruleSetName :: Lens' UpdateReceiptRule Text

The name of the receipt rule set that the receipt rule belongs to.

updateReceiptRule_rule :: Lens' UpdateReceiptRule ReceiptRule

A data structure that contains the updated receipt rule information.

Destructuring the Response

data UpdateReceiptRuleResponse

An empty element returned on a successful request.

See: newUpdateReceiptRuleResponse smart constructor.

Constructors

UpdateReceiptRuleResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateReceiptRuleResponse

Arguments

:: Int
$sel:httpStatus:UpdateReceiptRuleResponse'

-> UpdateReceiptRuleResponse

Create a value of UpdateReceiptRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateReceiptRuleResponse', updateReceiptRuleResponse_httpStatus - The response's http status code.

Response Lenses

updateReceiptRuleResponse_httpStatus :: Lens' UpdateReceiptRuleResponse Int

The response's http status code.