Network.AWS.SES.V2010_12_01.GetSendQuota

Description

Returns the user's current sending limits. This action is throttled at one request per second. POST HTTP1.1 Date: Thu, 18 Aug 2011 22:22:36 GMT Host: email.us-east-1.amazonaws.com Content-Type: application/x-www-form-urlencoded X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE, Signature=W1YdiNOtf0jN3t7Lv63qhz7UZc3RrcmQpkGbopvnj/Y=, Algorithm=HmacSHA256, SignedHeaders=Date;Host Content-Length: 94 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &Action=GetSendQuota &Timestamp=2011-08-18T22%3A22%3A36.000Z 127.0 200.0 1.0 273021c6-c866-11e0-b926-699e21c3af9e.

Documentation

data GetSendQuota

Constructors

GetSendQuota

data GetSendQuotaResponse

Constructors

GetSendQuotaResponse

Fields

_gsqrMax24HourSend :: Maybe Double

The maximum number of emails the user is allowed to send in a 24-hour interval.

_gsqrMaxSendRate :: Maybe Double

The maximum number of emails the user is allowed to send per second.

_gsqrSentLast24Hours :: Maybe Double

The number of emails sent during the previous 24 hours.

gsqrSentLast24Hours :: Lens' GetSendQuotaResponse (Maybe Double)

gsqrMaxSendRate :: Lens' GetSendQuotaResponse (Maybe Double)

gsqrMax24HourSend :: Lens' GetSendQuotaResponse (Maybe Double)