Network.AWS.SES.V2010_12_01.ListIdentities

Description

Returns a list containing all of the identities (email addresses and domains) for a specific AWS Account, regardless of verification status. This action is throttled at one request per second. POST HTTP1.1 Date: Sat, 12 May 2012 05:18:45 GMT Host: email.us-east-1.amazonaws.com Content-Type: application/x-www-form-urlencoded X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE, Signature=OruiFNV26DCZicLDaQmULHGbjbU8MbCc5aIoMMIuM=, Algorithm=HmacSHA256, SignedHeaders=Date;Host Content-Length: 115 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &Action=ListIdentities &Timestamp=2012-05-12T05%3A18%3A45.000Z& Version=2010-12-01 example.com user@example.com cacecf23-9bf1-11e1-9279-0100e8cf109a.

Documentation

listIdentities :: ListIdentities

Minimum specification for a ListIdentities request.

data ListIdentities

Constructors

ListIdentities

Fields

_lirIdentityType :: Maybe IdentityType

The type of the identities to list. Possible values are EmailAddress and Domain. If this parameter is omitted, then all identities will be listed.

_lirMaxItems :: Maybe Integer

The maximum number of identities per page. Possible values are 1-100 inclusive.

_lirNextToken :: Maybe Text

The token to use for pagination.

lirNextToken :: Lens' ListIdentities (Maybe Text)

lirMaxItems :: Lens' ListIdentities (Maybe Integer)

lirIdentityType :: Lens' ListIdentities (Maybe IdentityType)

data ListIdentitiesResponse

Constructors

ListIdentitiesResponse

Fields

_lisIdentities :: [Text]

A list of identities.

_lisNextToken :: Maybe Text

The token used for pagination.

lisNextToken :: Lens' ListIdentitiesResponse (Maybe Text)

lisIdentities :: Lens' ListIdentitiesResponse [Text]