Network.AWS.SES.V2010_12_01.ListVerifiedEmailAddresses

Description

Returns a list containing all of the email addresses that have been verified. The ListVerifiedEmailAddresses action is deprecated as of the May 15, 2012 release of Domain Verification. The ListIdentities action is now preferred. This action is throttled at one request per second. POST / HTTP/1.1 Date: Thu, 18 Aug 2011 22:05:09 GMT Host: email.us-east-1.amazonaws.com Content-Type: application/x-www-form-urlencoded X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE, Signature=II0+vvDKGMv71vToBwzR6vZ1hxe/VUE8tWEFUNTUqgE=, Algorithm=HmacSHA256, SignedHeaders=Date;Host Content-Length: 108 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &Action=ListVerifiedEmailAddresses &Timestamp=2011-08-18T22%3A05%3A09.000Z% example@amazon.com 3dd50e97-c865-11e0-b235-099eb63d928d.

Documentation

data ListVerifiedEmailAddresses

Constructors

ListVerifiedEmailAddresses

data ListVerifiedEmailAddressesResponse

Constructors

ListVerifiedEmailAddressesResponse

Fields

_lvearVerifiedEmailAddresses :: [Text]

A list of email addresses that have been verified.

lvearVerifiedEmailAddresses :: Iso' ListVerifiedEmailAddressesResponse [Text]