Network.AWS.SES.V2010_12_01.Types

Description

Amazon Simple Email Service (Amazon SES) is a cost-effective outbound-only email-sending service built on the reliable and scalable infrastructure that Amazon.com has developed to serve its own customer base. With Amazon SES, you can send transactional email, marketing messages, or any other type of high-quality content and you only pay for what you use. Along with high deliverability, Amazon SES provides easy, real-time access to your sending statistics and built-in notifications for bounces, complaints, and deliveries to help you fine-tune your email-sending strategy.

Documentation

data SES

Supported version (2010-12-01) of the Amazon Simple Email Service service.

xmlOptions :: Tagged a XMLOptions

data IdentityType

The type of the identities to list. Possible values are EmailAddress and Domain. If this parameter is omitted, then all identities will be listed.

Constructors

IdentityTypeDomain
Domain

IdentityTypeEmailAddress
EmailAddress

data NotificationType

The type of notifications that will be published to the specified Amazon SNS topic.

Constructors

NotificationTypeBounce
Bounce

NotificationTypeComplaint
Complaint

NotificationTypeDelivery
Delivery

data VerificationStatus

The verification status of the identity: Pending, Success, Failed, or TemporaryFailure.

Constructors

VerificationStatusFailed
Failed

VerificationStatusNotStarted
NotStarted

VerificationStatusPending
Pending

VerificationStatusSuccess
Success

VerificationStatusTemporaryFailure
TemporaryFailure

newtype RawMessage

The raw text of the message. The client is responsible for ensuring the following: Message must contain a header and a body, separated by a blank line. All required header fields must be present. Each part of a multipart MIME message must be formatted properly. MIME content types must be among those supported by Amazon SES. For more information, go to the Amazon SES Developer Guide. Content must be base64-encoded, if MIME requires it.

Constructors

RawMessage

Fields

_rmData :: ByteString

The raw data of the message. The client must ensure that the message format complies with Internet email standards regarding email header fields, MIME types, MIME encoding, and base64 encoding (if necessary). The To:, CC:, and BCC: headers in the raw message can contain a group list. For more information, go to the Amazon SES Developer Guide.

data Body

The message body.

Constructors

Body

Fields

_byText :: Maybe Content

The content of the message, in text format. Use this for text-based email clients, or clients on high-latency networks (such as mobile devices).

_byHtml :: Maybe Content

The content of the message, in HTML format. Use this for email clients that can process HTML. You can include clickable links, formatted text, and much more in an HTML message.

data Content

The subject of the message: A short summary of the content, which will appear in the recipient's inbox.

Constructors

Content

Fields

_ctData :: Text

The textual data of the content.

_ctCharset :: Maybe Text

The character set of the content.

data Destination

The destination for this email, composed of To:, CC:, and BCC: fields.

Constructors

Destination

Fields

_dnBccAddresses :: [Text]

The BCC: field(s) of the message.

_dnCcAddresses :: [Text]

The CC: field(s) of the message.

_dnToAddresses :: [Text]

The To: field(s) of the message.

data IdentityDkimAttributes

Represents the DKIM attributes of a verified email address or a domain.

Constructors

IdentityDkimAttributes

Fields

_idaDkimTokens :: [Text]

A set of character strings that represent the domain's identity. Using these tokens, you will need to create DNS CNAME records that point to DKIM public keys hosted by Amazon SES. Amazon Web Services will eventually detect that you have updated your DNS records; this detection process may take up to 72 hours. Upon successful detection, Amazon SES will be able to DKIM-sign email originating from that domain. (This only applies to domain identities, not email address identities.) For more information about creating DNS records using DKIM tokens, go to the Amazon SES Developer Guide.

_idaDkimEnabled :: Bool

True if DKIM signing is enabled for email sent from the identity; false otherwise.

_idaDkimVerificationStatus :: VerificationStatus

Describes whether Amazon SES has successfully verified the DKIM DNS records (tokens) published in the domain name's DNS. (This only applies to domain identities, not email address identities.).

data IdentityNotificationAttributes

Represents the notification attributes of an identity, including whether an identity has Amazon Simple Notification Service (Amazon SNS) topics set for bounce, complaint, and/or delivery notifications, and whether feedback forwarding is enabled for bounce and complaint notifications.

Constructors

IdentityNotificationAttributes

Fields

_inaDeliveryTopic :: Text

The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish delivery notifications.

_inaBounceTopic :: Text

The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish bounce notifications.

_inaComplaintTopic :: Text

The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish complaint notifications.

_inaForwardingEnabled :: Bool

Describes whether Amazon SES will forward bounce and complaint notifications as email. true indicates that Amazon SES will forward bounce and complaint notifications as email, while false indicates that bounce and complaint notifications will be published only to the specified bounce and complaint Amazon SNS topics.

data IdentityVerificationAttributes

Represents the verification attributes of a single identity.

Constructors

IdentityVerificationAttributes

Fields

_ivaVerificationToken :: Maybe Text

The verification token for a domain identity. Null for email address identities.

_ivaVerificationStatus :: VerificationStatus

The verification status of the identity: Pending, Success, Failed, or TemporaryFailure.

data Message

The message to be sent.

Constructors

Message

Fields

_meSubject :: Content

The subject of the message: A short summary of the content, which will appear in the recipient's inbox.

_meBody :: Body

The message body.

data SendDataPoint

Represents sending statistics data. Each SendDataPoint contains statistics for a 15-minute period of sending activity.

Constructors

SendDataPoint

Fields

_sdpRejects :: Maybe Integer

Number of emails rejected by Amazon SES.

_sdpComplaints :: Maybe Integer

Number of unwanted emails that were rejected by recipients.

_sdpDeliveryAttempts :: Maybe Integer

Number of emails that have been enqueued for sending.

_sdpBounces :: Maybe Integer

Number of emails that have bounced.

_sdpTimestamp :: Maybe ISO8601

Time of the data point.

rmData :: Iso' RawMessage ByteString

rawMessage :: Iso' ByteString RawMessage

byText :: Lens' Body (Maybe Content)

byHtml :: Lens' Body (Maybe Content)

ctData :: Lens' Content Text

ctCharset :: Lens' Content (Maybe Text)

dnToAddresses :: Lens' Destination [Text]

dnCcAddresses :: Lens' Destination [Text]

dnBccAddresses :: Lens' Destination [Text]

idaDkimVerificationStatus :: Lens' IdentityDkimAttributes VerificationStatus

idaDkimTokens :: Lens' IdentityDkimAttributes [Text]

idaDkimEnabled :: Lens' IdentityDkimAttributes Bool

inaForwardingEnabled :: Lens' IdentityNotificationAttributes Bool

inaDeliveryTopic :: Lens' IdentityNotificationAttributes Text

inaComplaintTopic :: Lens' IdentityNotificationAttributes Text

inaBounceTopic :: Lens' IdentityNotificationAttributes Text

ivaVerificationToken :: Lens' IdentityVerificationAttributes (Maybe Text)

ivaVerificationStatus :: Lens' IdentityVerificationAttributes VerificationStatus

meSubject :: Lens' Message Content

meBody :: Lens' Message Body

sdpTimestamp :: Lens' SendDataPoint (Maybe ISO8601)

sdpRejects :: Lens' SendDataPoint (Maybe Integer)

sdpDeliveryAttempts :: Lens' SendDataPoint (Maybe Integer)

sdpComplaints :: Lens' SendDataPoint (Maybe Integer)

sdpBounces :: Lens' SendDataPoint (Maybe Integer)