Network.AWS.SES.V2010_12_01.VerifyEmailAddress

Description

Verifies an email address. This action causes a confirmation email message to be sent to the specified address. The VerifyEmailAddress action is deprecated as of the May 15, 2012 release of Domain Verification. The VerifyEmailIdentity action is now preferred. This action is throttled at one request per second. POST HTTP1.1 Date: Thu, 18 Aug 2011 22:28:27 GMT Host: email.us-east-1.amazonaws.com Content-Type: application/x-www-form-urlencoded X-Amzn-Authorization: AWS3 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE, Signature=o9NK68jraFg5BnaTQiQhpxj2x1dGONOEFHHgsM6o5as=, Algorithm=HmacSHA256, SignedHeaders=Date;Host Content-Length: 132 AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &Action=VerifyEmailAddress &EmailAddress=user%40example.com &Timestamp=2011-08-18T22%3A28%3A27.000Z 8edd7eb2-c864-11e0-9f8f-3da8fc215a7e.

Documentation

data VerifyEmailAddress

Constructors

VerifyEmailAddress

Fields

_vearEmailAddress :: Text

The email address to be verified.

vearEmailAddress :: Iso' VerifyEmailAddress Text

data VerifyEmailAddressResponse

Constructors

VerifyEmailAddressResponse