Network.AWS.SES.VerifyDomainIdentity

Description

Adds a domain to the list of identities for your Amazon SES account in the current AWS Region and attempts to verify it. For more information about verifying domains, see Verifying Email Addresses and Domains in the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data VerifyDomainIdentity

Represents a request to begin Amazon SES domain verification and to generate the TXT records that you must publish to the DNS server of your domain to complete the verification. For information about domain verification, see the Amazon SES Developer Guide.

See: newVerifyDomainIdentity smart constructor.

Constructors

VerifyDomainIdentity'

Fields

domain :: Text

The domain to be verified.

newVerifyDomainIdentity

Arguments

:: Text
$sel:domain:VerifyDomainIdentity'

-> VerifyDomainIdentity

Create a value of VerifyDomainIdentity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domain:VerifyDomainIdentity', verifyDomainIdentity_domain - The domain to be verified.

Request Lenses

verifyDomainIdentity_domain :: Lens' VerifyDomainIdentity Text

The domain to be verified.

Destructuring the Response

data VerifyDomainIdentityResponse

Returns a TXT record that you must publish to the DNS server of your domain to complete domain verification with Amazon SES.

See: newVerifyDomainIdentityResponse smart constructor.

Constructors

VerifyDomainIdentityResponse'

Fields

httpStatus :: Int

The response's http status code.

verificationToken :: Text

A TXT record that you must place in the DNS settings of the domain to complete domain verification with Amazon SES.

As Amazon SES searches for the TXT record, the domain's verification status is "Pending". When Amazon SES detects the record, the domain's verification status changes to "Success". If Amazon SES is unable to detect the record within 72 hours, the domain's verification status changes to "Failed." In that case, if you still want to verify the domain, you must restart the verification process from the beginning.

newVerifyDomainIdentityResponse

Arguments

:: Int
$sel:httpStatus:VerifyDomainIdentityResponse'

-> Text
$sel:verificationToken:VerifyDomainIdentityResponse'

-> VerifyDomainIdentityResponse

Create a value of VerifyDomainIdentityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:VerifyDomainIdentityResponse', verifyDomainIdentityResponse_httpStatus - The response's http status code.

$sel:verificationToken:VerifyDomainIdentityResponse', verifyDomainIdentityResponse_verificationToken - A TXT record that you must place in the DNS settings of the domain to complete domain verification with Amazon SES.

As Amazon SES searches for the TXT record, the domain's verification status is "Pending". When Amazon SES detects the record, the domain's verification status changes to "Success". If Amazon SES is unable to detect the record within 72 hours, the domain's verification status changes to "Failed." In that case, if you still want to verify the domain, you must restart the verification process from the beginning.

Response Lenses

verifyDomainIdentityResponse_httpStatus :: Lens' VerifyDomainIdentityResponse Int

The response's http status code.

verifyDomainIdentityResponse_verificationToken :: Lens' VerifyDomainIdentityResponse Text

A TXT record that you must place in the DNS settings of the domain to complete domain verification with Amazon SES.

As Amazon SES searches for the TXT record, the domain's verification status is "Pending". When Amazon SES detects the record, the domain's verification status changes to "Success". If Amazon SES is unable to detect the record within 72 hours, the domain's verification status changes to "Failed." In that case, if you still want to verify the domain, you must restart the verification process from the beginning.