Network.AWS.SESV2.GetAccount

Description

Obtain information about the email-sending status and capabilities of your Amazon SES account in the current Amazon Web Services Region.

Creating a Request

data GetAccount

A request to obtain information about the email-sending capabilities of your Amazon SES account.

See: newGetAccount smart constructor.

Constructors

GetAccount'

newGetAccount :: GetAccount

Create a value of GetAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetAccountResponse

A list of details about the email-sending capabilities of your Amazon SES account in the current Amazon Web Services Region.

See: newGetAccountResponse smart constructor.

Constructors

GetAccountResponse'

Fields

enforcementStatus :: Maybe Text

The reputation status of your Amazon SES account. The status can be one of the following:

HEALTHY – There are no reputation-related issues that currently impact your account.

PROBATION – We've identified potential issues with your Amazon SES account. We're placing your account under review while you work on correcting these issues.

SHUTDOWN – Your account's ability to send email is currently paused because of an issue with the email sent from your account. When you correct the issue, you can contact us and request that your account's ability to send email is resumed.

dedicatedIpAutoWarmupEnabled :: Maybe Bool

Indicates whether or not the automatic warm-up feature is enabled for dedicated IP addresses that are associated with your account.

sendQuota :: Maybe SendQuota

An object that contains information about the per-day and per-second sending limits for your Amazon SES account in the current Amazon Web Services Region.

productionAccessEnabled :: Maybe Bool

Indicates whether or not your account has production access in the current Amazon Web Services Region.

If the value is false, then your account is in the sandbox. When your account is in the sandbox, you can only send email to verified identities. Additionally, the maximum number of emails you can send in a 24-hour period (your sending quota) is 200, and the maximum number of emails you can send per second (your maximum sending rate) is 1.

If the value is true, then your account has production access. When your account has production access, you can send email to any address. The sending quota and maximum sending rate for your account vary based on your specific use case.

details :: Maybe AccountDetails

An object that defines your account details.

suppressionAttributes :: Maybe SuppressionAttributes

An object that contains information about the email address suppression preferences for your account in the current Amazon Web Services Region.

sendingEnabled :: Maybe Bool

Indicates whether or not email sending is enabled for your Amazon SES account in the current Amazon Web Services Region.

httpStatus :: Int

The response's http status code.

newGetAccountResponse

Arguments

:: Int
$sel:httpStatus:GetAccountResponse'

-> GetAccountResponse

Create a value of GetAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enforcementStatus:GetAccountResponse', getAccountResponse_enforcementStatus - The reputation status of your Amazon SES account. The status can be one of the following:

HEALTHY – There are no reputation-related issues that currently impact your account.

PROBATION – We've identified potential issues with your Amazon SES account. We're placing your account under review while you work on correcting these issues.

SHUTDOWN – Your account's ability to send email is currently paused because of an issue with the email sent from your account. When you correct the issue, you can contact us and request that your account's ability to send email is resumed.

$sel:dedicatedIpAutoWarmupEnabled:GetAccountResponse', getAccountResponse_dedicatedIpAutoWarmupEnabled - Indicates whether or not the automatic warm-up feature is enabled for dedicated IP addresses that are associated with your account.

$sel:sendQuota:GetAccountResponse', getAccountResponse_sendQuota - An object that contains information about the per-day and per-second sending limits for your Amazon SES account in the current Amazon Web Services Region.

$sel:productionAccessEnabled:GetAccountResponse', getAccountResponse_productionAccessEnabled - Indicates whether or not your account has production access in the current Amazon Web Services Region.

If the value is false, then your account is in the sandbox. When your account is in the sandbox, you can only send email to verified identities. Additionally, the maximum number of emails you can send in a 24-hour period (your sending quota) is 200, and the maximum number of emails you can send per second (your maximum sending rate) is 1.

If the value is true, then your account has production access. When your account has production access, you can send email to any address. The sending quota and maximum sending rate for your account vary based on your specific use case.

$sel:details:GetAccountResponse', getAccountResponse_details - An object that defines your account details.

$sel:suppressionAttributes:GetAccountResponse', getAccountResponse_suppressionAttributes - An object that contains information about the email address suppression preferences for your account in the current Amazon Web Services Region.

$sel:sendingEnabled:GetAccountResponse', getAccountResponse_sendingEnabled - Indicates whether or not email sending is enabled for your Amazon SES account in the current Amazon Web Services Region.

$sel:httpStatus:GetAccountResponse', getAccountResponse_httpStatus - The response's http status code.

Response Lenses

getAccountResponse_enforcementStatus :: Lens' GetAccountResponse (Maybe Text)

The reputation status of your Amazon SES account. The status can be one of the following:

HEALTHY – There are no reputation-related issues that currently impact your account.

PROBATION – We've identified potential issues with your Amazon SES account. We're placing your account under review while you work on correcting these issues.

SHUTDOWN – Your account's ability to send email is currently paused because of an issue with the email sent from your account. When you correct the issue, you can contact us and request that your account's ability to send email is resumed.

getAccountResponse_dedicatedIpAutoWarmupEnabled :: Lens' GetAccountResponse (Maybe Bool)

Indicates whether or not the automatic warm-up feature is enabled for dedicated IP addresses that are associated with your account.

getAccountResponse_sendQuota :: Lens' GetAccountResponse (Maybe SendQuota)

An object that contains information about the per-day and per-second sending limits for your Amazon SES account in the current Amazon Web Services Region.

getAccountResponse_productionAccessEnabled :: Lens' GetAccountResponse (Maybe Bool)

Indicates whether or not your account has production access in the current Amazon Web Services Region.

If the value is false, then your account is in the sandbox. When your account is in the sandbox, you can only send email to verified identities. Additionally, the maximum number of emails you can send in a 24-hour period (your sending quota) is 200, and the maximum number of emails you can send per second (your maximum sending rate) is 1.

If the value is true, then your account has production access. When your account has production access, you can send email to any address. The sending quota and maximum sending rate for your account vary based on your specific use case.

getAccountResponse_details :: Lens' GetAccountResponse (Maybe AccountDetails)

An object that defines your account details.

getAccountResponse_suppressionAttributes :: Lens' GetAccountResponse (Maybe SuppressionAttributes)

An object that contains information about the email address suppression preferences for your account in the current Amazon Web Services Region.

getAccountResponse_sendingEnabled :: Lens' GetAccountResponse (Maybe Bool)

Indicates whether or not email sending is enabled for your Amazon SES account in the current Amazon Web Services Region.

getAccountResponse_httpStatus :: Lens' GetAccountResponse Int

The response's http status code.