Network.AWS.SESV2.ListCustomVerificationEmailTemplates

Description

Lists the existing custom verification email templates for your account in the current Amazon Web Services Region.

For more information about custom verification email templates, see Using Custom Verification Email Templates in the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data ListCustomVerificationEmailTemplates

Represents a request to list the existing custom verification email templates for your account.

See: newListCustomVerificationEmailTemplates smart constructor.

Constructors

ListCustomVerificationEmailTemplates'

Fields

nextToken :: Maybe Text

A token returned from a previous call to ListCustomVerificationEmailTemplates to indicate the position in the list of custom verification email templates.

pageSize :: Maybe Int

The number of results to show in a single call to ListCustomVerificationEmailTemplates. If the number of results is larger than the number you specified in this parameter, then the response includes a NextToken element, which you can use to obtain additional results.

The value you specify has to be at least 1, and can be no more than 50.

newListCustomVerificationEmailTemplates :: ListCustomVerificationEmailTemplates

Create a value of ListCustomVerificationEmailTemplates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListCustomVerificationEmailTemplates', listCustomVerificationEmailTemplates_nextToken - A token returned from a previous call to ListCustomVerificationEmailTemplates to indicate the position in the list of custom verification email templates.

$sel:pageSize:ListCustomVerificationEmailTemplates', listCustomVerificationEmailTemplates_pageSize - The number of results to show in a single call to ListCustomVerificationEmailTemplates. If the number of results is larger than the number you specified in this parameter, then the response includes a NextToken element, which you can use to obtain additional results.

The value you specify has to be at least 1, and can be no more than 50.

Request Lenses

listCustomVerificationEmailTemplates_nextToken :: Lens' ListCustomVerificationEmailTemplates (Maybe Text)

A token returned from a previous call to ListCustomVerificationEmailTemplates to indicate the position in the list of custom verification email templates.

listCustomVerificationEmailTemplates_pageSize :: Lens' ListCustomVerificationEmailTemplates (Maybe Int)

The number of results to show in a single call to ListCustomVerificationEmailTemplates. If the number of results is larger than the number you specified in this parameter, then the response includes a NextToken element, which you can use to obtain additional results.

The value you specify has to be at least 1, and can be no more than 50.

Destructuring the Response

data ListCustomVerificationEmailTemplatesResponse

The following elements are returned by the service.

See: newListCustomVerificationEmailTemplatesResponse smart constructor.

Constructors

ListCustomVerificationEmailTemplatesResponse'

Fields

nextToken :: Maybe Text

A token indicating that there are additional custom verification email templates available to be listed. Pass this token to a subsequent call to ListCustomVerificationEmailTemplates to retrieve the next 50 custom verification email templates.

customVerificationEmailTemplates :: Maybe [CustomVerificationEmailTemplateMetadata]

A list of the custom verification email templates that exist in your account.

httpStatus :: Int

The response's http status code.

newListCustomVerificationEmailTemplatesResponse

Arguments

:: Int
$sel:httpStatus:ListCustomVerificationEmailTemplatesResponse'

-> ListCustomVerificationEmailTemplatesResponse

Create a value of ListCustomVerificationEmailTemplatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListCustomVerificationEmailTemplates', listCustomVerificationEmailTemplatesResponse_nextToken - A token indicating that there are additional custom verification email templates available to be listed. Pass this token to a subsequent call to ListCustomVerificationEmailTemplates to retrieve the next 50 custom verification email templates.

$sel:customVerificationEmailTemplates:ListCustomVerificationEmailTemplatesResponse', listCustomVerificationEmailTemplatesResponse_customVerificationEmailTemplates - A list of the custom verification email templates that exist in your account.

$sel:httpStatus:ListCustomVerificationEmailTemplatesResponse', listCustomVerificationEmailTemplatesResponse_httpStatus - The response's http status code.

Response Lenses

listCustomVerificationEmailTemplatesResponse_nextToken :: Lens' ListCustomVerificationEmailTemplatesResponse (Maybe Text)

A token indicating that there are additional custom verification email templates available to be listed. Pass this token to a subsequent call to ListCustomVerificationEmailTemplates to retrieve the next 50 custom verification email templates.

listCustomVerificationEmailTemplatesResponse_customVerificationEmailTemplates :: Lens' ListCustomVerificationEmailTemplatesResponse (Maybe [CustomVerificationEmailTemplateMetadata])

A list of the custom verification email templates that exist in your account.

listCustomVerificationEmailTemplatesResponse_httpStatus :: Lens' ListCustomVerificationEmailTemplatesResponse Int

The response's http status code.