Network.AWS.SESV2.PutDeliverabilityDashboardOption

Description

Enable or disable the Deliverability dashboard. When you enable the Deliverability dashboard, you gain access to reputation, deliverability, and other metrics for the domains that you use to send email. You also gain the ability to perform predictive inbox placement tests.

When you use the Deliverability dashboard, you pay a monthly subscription charge, in addition to any other fees that you accrue by using Amazon SES and other Amazon Web Services services. For more information about the features and cost of a Deliverability dashboard subscription, see Amazon SES Pricing.

Creating a Request

data PutDeliverabilityDashboardOption

Enable or disable the Deliverability dashboard. When you enable the Deliverability dashboard, you gain access to reputation, deliverability, and other metrics for the domains that you use to send email using Amazon SES API v2. You also gain the ability to perform predictive inbox placement tests.

When you use the Deliverability dashboard, you pay a monthly subscription charge, in addition to any other fees that you accrue by using Amazon SES and other Amazon Web Services services. For more information about the features and cost of a Deliverability dashboard subscription, see Amazon Pinpoint Pricing.

See: newPutDeliverabilityDashboardOption smart constructor.

Constructors

PutDeliverabilityDashboardOption'

Fields

subscribedDomains :: Maybe [DomainDeliverabilityTrackingOption]

An array of objects, one for each verified domain that you use to send email and enabled the Deliverability dashboard for.

dashboardEnabled :: Bool

Specifies whether to enable the Deliverability dashboard. To enable the dashboard, set this value to true.

newPutDeliverabilityDashboardOption

Arguments

:: Bool
$sel:dashboardEnabled:PutDeliverabilityDashboardOption'

-> PutDeliverabilityDashboardOption

Create a value of PutDeliverabilityDashboardOption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscribedDomains:PutDeliverabilityDashboardOption', putDeliverabilityDashboardOption_subscribedDomains - An array of objects, one for each verified domain that you use to send email and enabled the Deliverability dashboard for.

$sel:dashboardEnabled:PutDeliverabilityDashboardOption', putDeliverabilityDashboardOption_dashboardEnabled - Specifies whether to enable the Deliverability dashboard. To enable the dashboard, set this value to true.

Request Lenses

putDeliverabilityDashboardOption_subscribedDomains :: Lens' PutDeliverabilityDashboardOption (Maybe [DomainDeliverabilityTrackingOption])

An array of objects, one for each verified domain that you use to send email and enabled the Deliverability dashboard for.

putDeliverabilityDashboardOption_dashboardEnabled :: Lens' PutDeliverabilityDashboardOption Bool

Specifies whether to enable the Deliverability dashboard. To enable the dashboard, set this value to true.

Destructuring the Response

data PutDeliverabilityDashboardOptionResponse

A response that indicates whether the Deliverability dashboard is enabled.

See: newPutDeliverabilityDashboardOptionResponse smart constructor.

Constructors

PutDeliverabilityDashboardOptionResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutDeliverabilityDashboardOptionResponse

Arguments

:: Int
$sel:httpStatus:PutDeliverabilityDashboardOptionResponse'

-> PutDeliverabilityDashboardOptionResponse

Create a value of PutDeliverabilityDashboardOptionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutDeliverabilityDashboardOptionResponse', putDeliverabilityDashboardOptionResponse_httpStatus - The response's http status code.

Response Lenses

putDeliverabilityDashboardOptionResponse_httpStatus :: Lens' PutDeliverabilityDashboardOptionResponse Int

The response's http status code.