Network.AWS.SESV2.Types.BulkEmailEntryResult

Description

Documentation

data BulkEmailEntryResult

The result of the SendBulkEmail operation of each specified BulkEmailEntry.

See: newBulkEmailEntryResult smart constructor.

Constructors

BulkEmailEntryResult'

Fields

status :: Maybe BulkEmailStatus

The status of a message sent using the SendBulkTemplatedEmail operation.

Possible values for this parameter include:

SUCCESS: Amazon SES accepted the message, and will attempt to deliver it to the recipients.

MESSAGE_REJECTED: The message was rejected because it contained a virus.

MAIL_FROM_DOMAIN_NOT_VERIFIED: The sender's email address or domain was not verified.

CONFIGURATION_SET_DOES_NOT_EXIST: The configuration set you specified does not exist.

TEMPLATE_DOES_NOT_EXIST: The template you specified does not exist.

ACCOUNT_SUSPENDED: Your account has been shut down because of issues related to your email sending practices.

ACCOUNT_THROTTLED: The number of emails you can send has been reduced because your account has exceeded its allocated sending limit.

ACCOUNT_DAILY_QUOTA_EXCEEDED: You have reached or exceeded the maximum number of emails you can send from your account in a 24-hour period.

INVALID_SENDING_POOL_NAME: The configuration set you specified refers to an IP pool that does not exist.

ACCOUNT_SENDING_PAUSED: Email sending for the Amazon SES account was disabled using the UpdateAccountSendingEnabled operation.

CONFIGURATION_SET_SENDING_PAUSED: Email sending for this configuration set was disabled using the UpdateConfigurationSetSendingEnabled operation.

INVALID_PARAMETER_VALUE: One or more of the parameters you specified when calling this operation was invalid. See the error message for additional information.

TRANSIENT_FAILURE: Amazon SES was unable to process your request because of a temporary issue.

FAILED: Amazon SES was unable to process your request. See the error message for additional information.

error :: Maybe Text

A description of an error that prevented a message being sent using the SendBulkTemplatedEmail operation.

messageId :: Maybe Text

The unique message identifier returned from the SendBulkTemplatedEmail operation.

newBulkEmailEntryResult :: BulkEmailEntryResult

Create a value of BulkEmailEntryResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:BulkEmailEntryResult', bulkEmailEntryResult_status - The status of a message sent using the SendBulkTemplatedEmail operation.

Possible values for this parameter include:

SUCCESS: Amazon SES accepted the message, and will attempt to deliver it to the recipients.

MESSAGE_REJECTED: The message was rejected because it contained a virus.

MAIL_FROM_DOMAIN_NOT_VERIFIED: The sender's email address or domain was not verified.

CONFIGURATION_SET_DOES_NOT_EXIST: The configuration set you specified does not exist.

TEMPLATE_DOES_NOT_EXIST: The template you specified does not exist.

ACCOUNT_SUSPENDED: Your account has been shut down because of issues related to your email sending practices.

ACCOUNT_THROTTLED: The number of emails you can send has been reduced because your account has exceeded its allocated sending limit.

ACCOUNT_DAILY_QUOTA_EXCEEDED: You have reached or exceeded the maximum number of emails you can send from your account in a 24-hour period.

INVALID_SENDING_POOL_NAME: The configuration set you specified refers to an IP pool that does not exist.

ACCOUNT_SENDING_PAUSED: Email sending for the Amazon SES account was disabled using the UpdateAccountSendingEnabled operation.

CONFIGURATION_SET_SENDING_PAUSED: Email sending for this configuration set was disabled using the UpdateConfigurationSetSendingEnabled operation.

INVALID_PARAMETER_VALUE: One or more of the parameters you specified when calling this operation was invalid. See the error message for additional information.

TRANSIENT_FAILURE: Amazon SES was unable to process your request because of a temporary issue.

FAILED: Amazon SES was unable to process your request. See the error message for additional information.

$sel:error:BulkEmailEntryResult', bulkEmailEntryResult_error - A description of an error that prevented a message being sent using the SendBulkTemplatedEmail operation.

$sel:messageId:BulkEmailEntryResult', bulkEmailEntryResult_messageId - The unique message identifier returned from the SendBulkTemplatedEmail operation.

bulkEmailEntryResult_status :: Lens' BulkEmailEntryResult (Maybe BulkEmailStatus)

The status of a message sent using the SendBulkTemplatedEmail operation.

Possible values for this parameter include:

SUCCESS: Amazon SES accepted the message, and will attempt to deliver it to the recipients.

MESSAGE_REJECTED: The message was rejected because it contained a virus.

MAIL_FROM_DOMAIN_NOT_VERIFIED: The sender's email address or domain was not verified.

CONFIGURATION_SET_DOES_NOT_EXIST: The configuration set you specified does not exist.

TEMPLATE_DOES_NOT_EXIST: The template you specified does not exist.

ACCOUNT_SUSPENDED: Your account has been shut down because of issues related to your email sending practices.

ACCOUNT_THROTTLED: The number of emails you can send has been reduced because your account has exceeded its allocated sending limit.

ACCOUNT_DAILY_QUOTA_EXCEEDED: You have reached or exceeded the maximum number of emails you can send from your account in a 24-hour period.

INVALID_SENDING_POOL_NAME: The configuration set you specified refers to an IP pool that does not exist.

ACCOUNT_SENDING_PAUSED: Email sending for the Amazon SES account was disabled using the UpdateAccountSendingEnabled operation.

CONFIGURATION_SET_SENDING_PAUSED: Email sending for this configuration set was disabled using the UpdateConfigurationSetSendingEnabled operation.

INVALID_PARAMETER_VALUE: One or more of the parameters you specified when calling this operation was invalid. See the error message for additional information.

TRANSIENT_FAILURE: Amazon SES was unable to process your request because of a temporary issue.

FAILED: Amazon SES was unable to process your request. See the error message for additional information.

bulkEmailEntryResult_error :: Lens' BulkEmailEntryResult (Maybe Text)

A description of an error that prevented a message being sent using the SendBulkTemplatedEmail operation.

bulkEmailEntryResult_messageId :: Lens' BulkEmailEntryResult (Maybe Text)

The unique message identifier returned from the SendBulkTemplatedEmail operation.