Network.AWS.SESV2.UpdateCustomVerificationEmailTemplate

Description

Updates an existing custom verification email template.

For more information about custom verification email templates, see Using Custom Verification Email Templates in the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data UpdateCustomVerificationEmailTemplate

Represents a request to update an existing custom verification email template.

See: newUpdateCustomVerificationEmailTemplate smart constructor.

Constructors

UpdateCustomVerificationEmailTemplate'

Fields

templateName :: Text

The name of the custom verification email template that you want to update.

fromEmailAddress :: Text

The email address that the custom verification email is sent from.

templateSubject :: Text

The subject line of the custom verification email.

templateContent :: Text

The content of the custom verification email. The total size of the email must be less than 10 MB. The message body may contain HTML, with some limitations. For more information, see Custom Verification Email Frequently Asked Questions in the Amazon SES Developer Guide.

successRedirectionURL :: Text

The URL that the recipient of the verification email is sent to if his or her address is successfully verified.

failureRedirectionURL :: Text

The URL that the recipient of the verification email is sent to if his or her address is not successfully verified.

newUpdateCustomVerificationEmailTemplate

Arguments

:: Text
$sel:templateName:UpdateCustomVerificationEmailTemplate'

-> Text
$sel:fromEmailAddress:UpdateCustomVerificationEmailTemplate'

-> Text
$sel:templateSubject:UpdateCustomVerificationEmailTemplate'

-> Text
$sel:templateContent:UpdateCustomVerificationEmailTemplate'

-> Text
$sel:successRedirectionURL:UpdateCustomVerificationEmailTemplate'

-> Text
$sel:failureRedirectionURL:UpdateCustomVerificationEmailTemplate'

-> UpdateCustomVerificationEmailTemplate

Create a value of UpdateCustomVerificationEmailTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:UpdateCustomVerificationEmailTemplate', updateCustomVerificationEmailTemplate_templateName - The name of the custom verification email template that you want to update.

$sel:fromEmailAddress:UpdateCustomVerificationEmailTemplate', updateCustomVerificationEmailTemplate_fromEmailAddress - The email address that the custom verification email is sent from.

$sel:templateSubject:UpdateCustomVerificationEmailTemplate', updateCustomVerificationEmailTemplate_templateSubject - The subject line of the custom verification email.

$sel:templateContent:UpdateCustomVerificationEmailTemplate', updateCustomVerificationEmailTemplate_templateContent - The content of the custom verification email. The total size of the email must be less than 10 MB. The message body may contain HTML, with some limitations. For more information, see Custom Verification Email Frequently Asked Questions in the Amazon SES Developer Guide.

$sel:successRedirectionURL:UpdateCustomVerificationEmailTemplate', updateCustomVerificationEmailTemplate_successRedirectionURL - The URL that the recipient of the verification email is sent to if his or her address is successfully verified.

$sel:failureRedirectionURL:UpdateCustomVerificationEmailTemplate', updateCustomVerificationEmailTemplate_failureRedirectionURL - The URL that the recipient of the verification email is sent to if his or her address is not successfully verified.

Request Lenses

updateCustomVerificationEmailTemplate_templateName :: Lens' UpdateCustomVerificationEmailTemplate Text

The name of the custom verification email template that you want to update.

updateCustomVerificationEmailTemplate_fromEmailAddress :: Lens' UpdateCustomVerificationEmailTemplate Text

The email address that the custom verification email is sent from.

updateCustomVerificationEmailTemplate_templateSubject :: Lens' UpdateCustomVerificationEmailTemplate Text

The subject line of the custom verification email.

updateCustomVerificationEmailTemplate_templateContent :: Lens' UpdateCustomVerificationEmailTemplate Text

The content of the custom verification email. The total size of the email must be less than 10 MB. The message body may contain HTML, with some limitations. For more information, see Custom Verification Email Frequently Asked Questions in the Amazon SES Developer Guide.

updateCustomVerificationEmailTemplate_successRedirectionURL :: Lens' UpdateCustomVerificationEmailTemplate Text

The URL that the recipient of the verification email is sent to if his or her address is successfully verified.

updateCustomVerificationEmailTemplate_failureRedirectionURL :: Lens' UpdateCustomVerificationEmailTemplate Text

The URL that the recipient of the verification email is sent to if his or her address is not successfully verified.

Destructuring the Response

data UpdateCustomVerificationEmailTemplateResponse

If the action is successful, the service sends back an HTTP 200 response with an empty HTTP body.

See: newUpdateCustomVerificationEmailTemplateResponse smart constructor.

Constructors

UpdateCustomVerificationEmailTemplateResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateCustomVerificationEmailTemplateResponse

Arguments

:: Int
$sel:httpStatus:UpdateCustomVerificationEmailTemplateResponse'

-> UpdateCustomVerificationEmailTemplateResponse

Create a value of UpdateCustomVerificationEmailTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateCustomVerificationEmailTemplateResponse', updateCustomVerificationEmailTemplateResponse_httpStatus - The response's http status code.

Response Lenses

updateCustomVerificationEmailTemplateResponse_httpStatus :: Lens' UpdateCustomVerificationEmailTemplateResponse Int

The response's http status code.