Network.AWS.SESv2.CreateConfigurationSet

Description

Create a configuration set. Configuration sets are groups of rules that you can apply to the emails that you send. You apply a configuration set to an email by specifying the name of the configuration set when you call the Amazon SES API v2. When you apply a configuration set to an email, all of the rules in that configuration set are applied to the email.

Creating a Request

data CreateConfigurationSet

A request to create a configuration set.

See: newCreateConfigurationSet smart constructor.

Constructors

CreateConfigurationSet'

Fields

sendingOptions :: Maybe SendingOptions

An object that defines whether or not Amazon SES can send email that you send using the configuration set.

deliveryOptions :: Maybe DeliveryOptions

An object that defines the dedicated IP pool that is used to send emails that you send using the configuration set.

trackingOptions :: Maybe TrackingOptions

An object that defines the open and click tracking options for emails that you send using the configuration set.

reputationOptions :: Maybe ReputationOptions

An object that defines whether or not Amazon SES collects reputation metrics for the emails that you send that use the configuration set.

tags :: Maybe [Tag]

An array of objects that define the tags (keys and values) to associate with the configuration set.

suppressionOptions :: Maybe SuppressionOptions

configurationSetName :: Text

The name of the configuration set. The name can contain up to 64 alphanumeric characters, including letters, numbers, hyphens (-) and underscores (_) only.

newCreateConfigurationSet

Arguments

:: Text
$sel:configurationSetName:CreateConfigurationSet'

-> CreateConfigurationSet

Create a value of CreateConfigurationSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sendingOptions:CreateConfigurationSet', createConfigurationSet_sendingOptions - An object that defines whether or not Amazon SES can send email that you send using the configuration set.

$sel:deliveryOptions:CreateConfigurationSet', createConfigurationSet_deliveryOptions - An object that defines the dedicated IP pool that is used to send emails that you send using the configuration set.

$sel:trackingOptions:CreateConfigurationSet', createConfigurationSet_trackingOptions - An object that defines the open and click tracking options for emails that you send using the configuration set.

$sel:reputationOptions:CreateConfigurationSet', createConfigurationSet_reputationOptions - An object that defines whether or not Amazon SES collects reputation metrics for the emails that you send that use the configuration set.

$sel:tags:CreateConfigurationSet', createConfigurationSet_tags - An array of objects that define the tags (keys and values) to associate with the configuration set.

$sel:suppressionOptions:CreateConfigurationSet', createConfigurationSet_suppressionOptions - Undocumented member.

$sel:configurationSetName:CreateConfigurationSet', createConfigurationSet_configurationSetName - The name of the configuration set. The name can contain up to 64 alphanumeric characters, including letters, numbers, hyphens (-) and underscores (_) only.

Request Lenses

createConfigurationSet_sendingOptions :: Lens' CreateConfigurationSet (Maybe SendingOptions)

An object that defines whether or not Amazon SES can send email that you send using the configuration set.

createConfigurationSet_deliveryOptions :: Lens' CreateConfigurationSet (Maybe DeliveryOptions)

An object that defines the dedicated IP pool that is used to send emails that you send using the configuration set.

createConfigurationSet_trackingOptions :: Lens' CreateConfigurationSet (Maybe TrackingOptions)

An object that defines the open and click tracking options for emails that you send using the configuration set.

createConfigurationSet_reputationOptions :: Lens' CreateConfigurationSet (Maybe ReputationOptions)

An object that defines whether or not Amazon SES collects reputation metrics for the emails that you send that use the configuration set.

createConfigurationSet_tags :: Lens' CreateConfigurationSet (Maybe [Tag])

An array of objects that define the tags (keys and values) to associate with the configuration set.

createConfigurationSet_suppressionOptions :: Lens' CreateConfigurationSet (Maybe SuppressionOptions)

Undocumented member.

createConfigurationSet_configurationSetName :: Lens' CreateConfigurationSet Text

The name of the configuration set. The name can contain up to 64 alphanumeric characters, including letters, numbers, hyphens (-) and underscores (_) only.

Destructuring the Response

data CreateConfigurationSetResponse

An HTTP 200 response if the request succeeds, or an error message if the request fails.

See: newCreateConfigurationSetResponse smart constructor.

Constructors

CreateConfigurationSetResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateConfigurationSetResponse

Arguments

:: Int
$sel:httpStatus:CreateConfigurationSetResponse'

-> CreateConfigurationSetResponse

Create a value of CreateConfigurationSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateConfigurationSetResponse', createConfigurationSetResponse_httpStatus - The response's http status code.

Response Lenses

createConfigurationSetResponse_httpStatus :: Lens' CreateConfigurationSetResponse Int

The response's http status code.