Network.AWS.SESv2.CreateContactList

Description

Creates a contact list.

Creating a Request

data CreateContactList

See: newCreateContactList smart constructor.

Constructors

CreateContactList'

Fields

topics :: Maybe [Topic]

An interest group, theme, or label within a list. A contact list can have multiple topics.

description :: Maybe Text

A description of what the contact list is about.

tags :: Maybe [Tag]

The tags associated with a contact list.

contactListName :: Text

The name of the contact list.

newCreateContactList

Arguments

:: Text
$sel:contactListName:CreateContactList'

-> CreateContactList

Create a value of CreateContactList with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:topics:CreateContactList', createContactList_topics - An interest group, theme, or label within a list. A contact list can have multiple topics.

$sel:description:CreateContactList', createContactList_description - A description of what the contact list is about.

$sel:tags:CreateContactList', createContactList_tags - The tags associated with a contact list.

$sel:contactListName:CreateContactList', createContactList_contactListName - The name of the contact list.

Request Lenses

createContactList_topics :: Lens' CreateContactList (Maybe [Topic])

An interest group, theme, or label within a list. A contact list can have multiple topics.

createContactList_description :: Lens' CreateContactList (Maybe Text)

A description of what the contact list is about.

createContactList_tags :: Lens' CreateContactList (Maybe [Tag])

The tags associated with a contact list.

createContactList_contactListName :: Lens' CreateContactList Text

The name of the contact list.

Destructuring the Response

data CreateContactListResponse

See: newCreateContactListResponse smart constructor.

Constructors

CreateContactListResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateContactListResponse

Arguments

:: Int
$sel:httpStatus:CreateContactListResponse'

-> CreateContactListResponse

Create a value of CreateContactListResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateContactListResponse', createContactListResponse_httpStatus - The response's http status code.

Response Lenses

createContactListResponse_httpStatus :: Lens' CreateContactListResponse Int

The response's http status code.