Network.AWS.SESv2.CreateCustomVerificationEmailTemplate

Description

Creates a new custom verification email template.

For more information about custom verification email templates, see Using Custom Verification Email Templates in the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data CreateCustomVerificationEmailTemplate

Represents a request to create a custom verification email template.

See: newCreateCustomVerificationEmailTemplate smart constructor.

Constructors

CreateCustomVerificationEmailTemplate'

Fields

templateName :: Text

The name of the custom verification email template.

fromEmailAddress :: Text

The email address that the custom verification email is sent from.

templateSubject :: Text

The subject line of the custom verification email.

templateContent :: Text

The content of the custom verification email. The total size of the email must be less than 10 MB. The message body may contain HTML, with some limitations. For more information, see Custom Verification Email Frequently Asked Questions in the Amazon SES Developer Guide.

successRedirectionURL :: Text

The URL that the recipient of the verification email is sent to if his or her address is successfully verified.

failureRedirectionURL :: Text

The URL that the recipient of the verification email is sent to if his or her address is not successfully verified.

newCreateCustomVerificationEmailTemplate

Arguments

:: Text
$sel:templateName:CreateCustomVerificationEmailTemplate'

-> Text
$sel:fromEmailAddress:CreateCustomVerificationEmailTemplate'

-> Text
$sel:templateSubject:CreateCustomVerificationEmailTemplate'

-> Text
$sel:templateContent:CreateCustomVerificationEmailTemplate'

-> Text
$sel:successRedirectionURL:CreateCustomVerificationEmailTemplate'

-> Text
$sel:failureRedirectionURL:CreateCustomVerificationEmailTemplate'

-> CreateCustomVerificationEmailTemplate

Create a value of CreateCustomVerificationEmailTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:CreateCustomVerificationEmailTemplate', createCustomVerificationEmailTemplate_templateName - The name of the custom verification email template.

$sel:fromEmailAddress:CreateCustomVerificationEmailTemplate', createCustomVerificationEmailTemplate_fromEmailAddress - The email address that the custom verification email is sent from.

$sel:templateSubject:CreateCustomVerificationEmailTemplate', createCustomVerificationEmailTemplate_templateSubject - The subject line of the custom verification email.

$sel:templateContent:CreateCustomVerificationEmailTemplate', createCustomVerificationEmailTemplate_templateContent - The content of the custom verification email. The total size of the email must be less than 10 MB. The message body may contain HTML, with some limitations. For more information, see Custom Verification Email Frequently Asked Questions in the Amazon SES Developer Guide.

$sel:successRedirectionURL:CreateCustomVerificationEmailTemplate', createCustomVerificationEmailTemplate_successRedirectionURL - The URL that the recipient of the verification email is sent to if his or her address is successfully verified.

$sel:failureRedirectionURL:CreateCustomVerificationEmailTemplate', createCustomVerificationEmailTemplate_failureRedirectionURL - The URL that the recipient of the verification email is sent to if his or her address is not successfully verified.

Request Lenses

createCustomVerificationEmailTemplate_templateName :: Lens' CreateCustomVerificationEmailTemplate Text

The name of the custom verification email template.

createCustomVerificationEmailTemplate_fromEmailAddress :: Lens' CreateCustomVerificationEmailTemplate Text

The email address that the custom verification email is sent from.

createCustomVerificationEmailTemplate_templateSubject :: Lens' CreateCustomVerificationEmailTemplate Text

The subject line of the custom verification email.

createCustomVerificationEmailTemplate_templateContent :: Lens' CreateCustomVerificationEmailTemplate Text

The content of the custom verification email. The total size of the email must be less than 10 MB. The message body may contain HTML, with some limitations. For more information, see Custom Verification Email Frequently Asked Questions in the Amazon SES Developer Guide.

createCustomVerificationEmailTemplate_successRedirectionURL :: Lens' CreateCustomVerificationEmailTemplate Text

The URL that the recipient of the verification email is sent to if his or her address is successfully verified.

createCustomVerificationEmailTemplate_failureRedirectionURL :: Lens' CreateCustomVerificationEmailTemplate Text

The URL that the recipient of the verification email is sent to if his or her address is not successfully verified.

Destructuring the Response

data CreateCustomVerificationEmailTemplateResponse

If the action is successful, the service sends back an HTTP 200 response with an empty HTTP body.

See: newCreateCustomVerificationEmailTemplateResponse smart constructor.

Constructors

CreateCustomVerificationEmailTemplateResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateCustomVerificationEmailTemplateResponse

Arguments

:: Int
$sel:httpStatus:CreateCustomVerificationEmailTemplateResponse'

-> CreateCustomVerificationEmailTemplateResponse

Create a value of CreateCustomVerificationEmailTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateCustomVerificationEmailTemplateResponse', createCustomVerificationEmailTemplateResponse_httpStatus - The response's http status code.

Response Lenses

createCustomVerificationEmailTemplateResponse_httpStatus :: Lens' CreateCustomVerificationEmailTemplateResponse Int

The response's http status code.