Network.AWS.SESv2.CreateEmailIdentity

Description

Starts the process of verifying an email identity. An identity is an email address or domain that you use when you send email. Before you can use an identity to send email, you first have to verify it. By verifying an identity, you demonstrate that you're the owner of the identity, and that you've given Amazon SES API v2 permission to send email from the identity.

When you verify an email address, Amazon SES sends an email to the address. Your email address is verified as soon as you follow the link in the verification email.

When you verify a domain without specifying the DkimSigningAttributes object, this operation provides a set of DKIM tokens. You can convert these tokens into CNAME records, which you then add to the DNS configuration for your domain. Your domain is verified when Amazon SES detects these records in the DNS configuration for your domain. This verification method is known as Easy DKIM.

Alternatively, you can perform the verification process by providing your own public-private key pair. This verification method is known as Bring Your Own DKIM (BYODKIM). To use BYODKIM, your call to the CreateEmailIdentity operation has to include the DkimSigningAttributes object. When you specify this object, you provide a selector (a component of the DNS record name that identifies the public key to use for DKIM authentication) and a private key.

When you verify a domain, this operation provides a set of DKIM tokens, which you can convert into CNAME tokens. You add these CNAME tokens to the DNS configuration for your domain. Your domain is verified when Amazon SES detects these records in the DNS configuration for your domain. For some DNS providers, it can take 72 hours or more to complete the domain verification process.

Additionally, you can associate an existing configuration set with the email identity that you're verifying.

Creating a Request

data CreateEmailIdentity

A request to begin the verification process for an email identity (an email address or domain).

See: newCreateEmailIdentity smart constructor.

Constructors

CreateEmailIdentity'

Fields

configurationSetName :: Maybe Text

The configuration set to use by default when sending from this identity. Note that any configuration set defined in the email sending request takes precedence.

dkimSigningAttributes :: Maybe DkimSigningAttributes

If your request includes this object, Amazon SES configures the identity to use Bring Your Own DKIM (BYODKIM) for DKIM authentication purposes, or, configures the key length to be used for Easy DKIM.

You can only specify this object if the email identity is a domain, as opposed to an address.

tags :: Maybe [Tag]

An array of objects that define the tags (keys and values) to associate with the email identity.

emailIdentity :: Text

The email address or domain to verify.

newCreateEmailIdentity

Arguments

:: Text
$sel:emailIdentity:CreateEmailIdentity'

-> CreateEmailIdentity

Create a value of CreateEmailIdentity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationSetName:CreateEmailIdentity', createEmailIdentity_configurationSetName - The configuration set to use by default when sending from this identity. Note that any configuration set defined in the email sending request takes precedence.

$sel:dkimSigningAttributes:CreateEmailIdentity', createEmailIdentity_dkimSigningAttributes - If your request includes this object, Amazon SES configures the identity to use Bring Your Own DKIM (BYODKIM) for DKIM authentication purposes, or, configures the key length to be used for Easy DKIM.

You can only specify this object if the email identity is a domain, as opposed to an address.

$sel:tags:CreateEmailIdentity', createEmailIdentity_tags - An array of objects that define the tags (keys and values) to associate with the email identity.

$sel:emailIdentity:CreateEmailIdentity', createEmailIdentity_emailIdentity - The email address or domain to verify.

Request Lenses

createEmailIdentity_configurationSetName :: Lens' CreateEmailIdentity (Maybe Text)

The configuration set to use by default when sending from this identity. Note that any configuration set defined in the email sending request takes precedence.

createEmailIdentity_dkimSigningAttributes :: Lens' CreateEmailIdentity (Maybe DkimSigningAttributes)

If your request includes this object, Amazon SES configures the identity to use Bring Your Own DKIM (BYODKIM) for DKIM authentication purposes, or, configures the key length to be used for Easy DKIM.

You can only specify this object if the email identity is a domain, as opposed to an address.

createEmailIdentity_tags :: Lens' CreateEmailIdentity (Maybe [Tag])

An array of objects that define the tags (keys and values) to associate with the email identity.

createEmailIdentity_emailIdentity :: Lens' CreateEmailIdentity Text

The email address or domain to verify.

Destructuring the Response

data CreateEmailIdentityResponse

If the email identity is a domain, this object contains information about the DKIM verification status for the domain.

If the email identity is an email address, this object is empty.

See: newCreateEmailIdentityResponse smart constructor.

Constructors

CreateEmailIdentityResponse'

Fields

dkimAttributes :: Maybe DkimAttributes

An object that contains information about the DKIM attributes for the identity.

verifiedForSendingStatus :: Maybe Bool

Specifies whether or not the identity is verified. You can only send email from verified email addresses or domains. For more information about verifying identities, see the Amazon Pinpoint User Guide.

identityType :: Maybe IdentityType

The email identity type. Note: the MANAGED_DOMAIN identity type is not supported.

httpStatus :: Int

The response's http status code.

newCreateEmailIdentityResponse

Arguments

:: Int
$sel:httpStatus:CreateEmailIdentityResponse'

-> CreateEmailIdentityResponse

Create a value of CreateEmailIdentityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dkimAttributes:CreateEmailIdentityResponse', createEmailIdentityResponse_dkimAttributes - An object that contains information about the DKIM attributes for the identity.

$sel:verifiedForSendingStatus:CreateEmailIdentityResponse', createEmailIdentityResponse_verifiedForSendingStatus - Specifies whether or not the identity is verified. You can only send email from verified email addresses or domains. For more information about verifying identities, see the Amazon Pinpoint User Guide.

$sel:identityType:CreateEmailIdentityResponse', createEmailIdentityResponse_identityType - The email identity type. Note: the MANAGED_DOMAIN identity type is not supported.

$sel:httpStatus:CreateEmailIdentityResponse', createEmailIdentityResponse_httpStatus - The response's http status code.

Response Lenses

createEmailIdentityResponse_dkimAttributes :: Lens' CreateEmailIdentityResponse (Maybe DkimAttributes)

An object that contains information about the DKIM attributes for the identity.

createEmailIdentityResponse_verifiedForSendingStatus :: Lens' CreateEmailIdentityResponse (Maybe Bool)

Specifies whether or not the identity is verified. You can only send email from verified email addresses or domains. For more information about verifying identities, see the Amazon Pinpoint User Guide.

createEmailIdentityResponse_identityType :: Lens' CreateEmailIdentityResponse (Maybe IdentityType)

The email identity type. Note: the MANAGED_DOMAIN identity type is not supported.

createEmailIdentityResponse_httpStatus :: Lens' CreateEmailIdentityResponse Int

The response's http status code.