Network.AWS.SESv2.DeleteEmailTemplate

Description

Deletes an email template.

You can execute this operation no more than once per second.

Creating a Request

data DeleteEmailTemplate

Represents a request to delete an email template. For more information, see the Amazon SES Developer Guide.

See: newDeleteEmailTemplate smart constructor.

Constructors

DeleteEmailTemplate'

Fields

templateName :: Text

The name of the template to be deleted.

newDeleteEmailTemplate

Arguments

:: Text
$sel:templateName:DeleteEmailTemplate'

-> DeleteEmailTemplate

Create a value of DeleteEmailTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:DeleteEmailTemplate', deleteEmailTemplate_templateName - The name of the template to be deleted.

Request Lenses

deleteEmailTemplate_templateName :: Lens' DeleteEmailTemplate Text

The name of the template to be deleted.

Destructuring the Response

data DeleteEmailTemplateResponse

If the action is successful, the service sends back an HTTP 200 response with an empty HTTP body.

See: newDeleteEmailTemplateResponse smart constructor.

Constructors

DeleteEmailTemplateResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteEmailTemplateResponse

Arguments

:: Int
$sel:httpStatus:DeleteEmailTemplateResponse'

-> DeleteEmailTemplateResponse

Create a value of DeleteEmailTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteEmailTemplateResponse', deleteEmailTemplateResponse_httpStatus - The response's http status code.

Response Lenses

deleteEmailTemplateResponse_httpStatus :: Lens' DeleteEmailTemplateResponse Int

The response's http status code.