Network.AWS.SESv2.GetDedicatedIps

Description

List the dedicated IP addresses that are associated with your Amazon Web Services account.

Creating a Request

data GetDedicatedIps

A request to obtain more information about dedicated IP pools.

See: newGetDedicatedIps smart constructor.

Constructors

GetDedicatedIps'

Fields

poolName :: Maybe Text

The name of the IP pool that the dedicated IP address is associated with.

nextToken :: Maybe Text

A token returned from a previous call to GetDedicatedIps to indicate the position of the dedicated IP pool in the list of IP pools.

pageSize :: Maybe Int

The number of results to show in a single call to GetDedicatedIpsRequest. If the number of results is larger than the number you specified in this parameter, then the response includes a NextToken element, which you can use to obtain additional results.

newGetDedicatedIps :: GetDedicatedIps

Create a value of GetDedicatedIps with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:poolName:GetDedicatedIps', getDedicatedIps_poolName - The name of the IP pool that the dedicated IP address is associated with.

$sel:nextToken:GetDedicatedIps', getDedicatedIps_nextToken - A token returned from a previous call to GetDedicatedIps to indicate the position of the dedicated IP pool in the list of IP pools.

$sel:pageSize:GetDedicatedIps', getDedicatedIps_pageSize - The number of results to show in a single call to GetDedicatedIpsRequest. If the number of results is larger than the number you specified in this parameter, then the response includes a NextToken element, which you can use to obtain additional results.

Request Lenses

getDedicatedIps_poolName :: Lens' GetDedicatedIps (Maybe Text)

The name of the IP pool that the dedicated IP address is associated with.

getDedicatedIps_nextToken :: Lens' GetDedicatedIps (Maybe Text)

A token returned from a previous call to GetDedicatedIps to indicate the position of the dedicated IP pool in the list of IP pools.

getDedicatedIps_pageSize :: Lens' GetDedicatedIps (Maybe Int)

The number of results to show in a single call to GetDedicatedIpsRequest. If the number of results is larger than the number you specified in this parameter, then the response includes a NextToken element, which you can use to obtain additional results.

Destructuring the Response

data GetDedicatedIpsResponse

Information about the dedicated IP addresses that are associated with your Amazon Web Services account.

See: newGetDedicatedIpsResponse smart constructor.

Constructors

GetDedicatedIpsResponse'

Fields

nextToken :: Maybe Text

A token that indicates that there are additional dedicated IP addresses to list. To view additional addresses, issue another request to GetDedicatedIps, passing this token in the NextToken parameter.

dedicatedIps :: Maybe [DedicatedIp]

A list of dedicated IP addresses that are associated with your Amazon Web Services account.

httpStatus :: Int

The response's http status code.

newGetDedicatedIpsResponse

Arguments

:: Int
$sel:httpStatus:GetDedicatedIpsResponse'

-> GetDedicatedIpsResponse

Create a value of GetDedicatedIpsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetDedicatedIps', getDedicatedIpsResponse_nextToken - A token that indicates that there are additional dedicated IP addresses to list. To view additional addresses, issue another request to GetDedicatedIps, passing this token in the NextToken parameter.

$sel:dedicatedIps:GetDedicatedIpsResponse', getDedicatedIpsResponse_dedicatedIps - A list of dedicated IP addresses that are associated with your Amazon Web Services account.

$sel:httpStatus:GetDedicatedIpsResponse', getDedicatedIpsResponse_httpStatus - The response's http status code.

Response Lenses

getDedicatedIpsResponse_nextToken :: Lens' GetDedicatedIpsResponse (Maybe Text)

A token that indicates that there are additional dedicated IP addresses to list. To view additional addresses, issue another request to GetDedicatedIps, passing this token in the NextToken parameter.

getDedicatedIpsResponse_dedicatedIps :: Lens' GetDedicatedIpsResponse (Maybe [DedicatedIp])

A list of dedicated IP addresses that are associated with your Amazon Web Services account.

getDedicatedIpsResponse_httpStatus :: Lens' GetDedicatedIpsResponse Int

The response's http status code.