Network.AWS.SESv2.GetDeliverabilityDashboardOptions

Description

Retrieve information about the status of the Deliverability dashboard for your account. When the Deliverability dashboard is enabled, you gain access to reputation, deliverability, and other metrics for the domains that you use to send email. You also gain the ability to perform predictive inbox placement tests.

When you use the Deliverability dashboard, you pay a monthly subscription charge, in addition to any other fees that you accrue by using Amazon SES and other AWS services. For more information about the features and cost of a Deliverability dashboard subscription, see Amazon SES Pricing.

Creating a Request

data GetDeliverabilityDashboardOptions

Retrieve information about the status of the Deliverability dashboard for your AWS account. When the Deliverability dashboard is enabled, you gain access to reputation, deliverability, and other metrics for your domains. You also gain the ability to perform predictive inbox placement tests.

When you use the Deliverability dashboard, you pay a monthly subscription charge, in addition to any other fees that you accrue by using Amazon SES and other AWS services. For more information about the features and cost of a Deliverability dashboard subscription, see Amazon Pinpoint Pricing.

See: newGetDeliverabilityDashboardOptions smart constructor.

Constructors

GetDeliverabilityDashboardOptions'

newGetDeliverabilityDashboardOptions :: GetDeliverabilityDashboardOptions

Create a value of GetDeliverabilityDashboardOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetDeliverabilityDashboardOptionsResponse

An object that shows the status of the Deliverability dashboard.

See: newGetDeliverabilityDashboardOptionsResponse smart constructor.

Constructors

GetDeliverabilityDashboardOptionsResponse'

Fields

subscriptionExpiryDate :: Maybe POSIX

The date, in Unix time format, when your current subscription to the Deliverability dashboard is scheduled to expire, if your subscription is scheduled to expire at the end of the current calendar month. This value is null if you have an active subscription that isn’t due to expire at the end of the month.

accountStatus :: Maybe DeliverabilityDashboardAccountStatus

The current status of your Deliverability dashboard subscription. If this value is PENDING_EXPIRATION, your subscription is scheduled to expire at the end of the current calendar month.

activeSubscribedDomains :: Maybe [DomainDeliverabilityTrackingOption]

An array of objects, one for each verified domain that you use to send email and currently has an active Deliverability dashboard subscription that isn’t scheduled to expire at the end of the current calendar month.

pendingExpirationSubscribedDomains :: Maybe [DomainDeliverabilityTrackingOption]

An array of objects, one for each verified domain that you use to send email and currently has an active Deliverability dashboard subscription that's scheduled to expire at the end of the current calendar month.

httpStatus :: Int

The response's http status code.

dashboardEnabled :: Bool

Specifies whether the Deliverability dashboard is enabled. If this value is true, the dashboard is enabled.

newGetDeliverabilityDashboardOptionsResponse

Arguments

:: Int
$sel:httpStatus:GetDeliverabilityDashboardOptionsResponse'

-> Bool
$sel:dashboardEnabled:GetDeliverabilityDashboardOptionsResponse'

-> GetDeliverabilityDashboardOptionsResponse

Create a value of GetDeliverabilityDashboardOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscriptionExpiryDate:GetDeliverabilityDashboardOptionsResponse', getDeliverabilityDashboardOptionsResponse_subscriptionExpiryDate - The date, in Unix time format, when your current subscription to the Deliverability dashboard is scheduled to expire, if your subscription is scheduled to expire at the end of the current calendar month. This value is null if you have an active subscription that isn’t due to expire at the end of the month.

$sel:accountStatus:GetDeliverabilityDashboardOptionsResponse', getDeliverabilityDashboardOptionsResponse_accountStatus - The current status of your Deliverability dashboard subscription. If this value is PENDING_EXPIRATION, your subscription is scheduled to expire at the end of the current calendar month.

$sel:activeSubscribedDomains:GetDeliverabilityDashboardOptionsResponse', getDeliverabilityDashboardOptionsResponse_activeSubscribedDomains - An array of objects, one for each verified domain that you use to send email and currently has an active Deliverability dashboard subscription that isn’t scheduled to expire at the end of the current calendar month.

$sel:pendingExpirationSubscribedDomains:GetDeliverabilityDashboardOptionsResponse', getDeliverabilityDashboardOptionsResponse_pendingExpirationSubscribedDomains - An array of objects, one for each verified domain that you use to send email and currently has an active Deliverability dashboard subscription that's scheduled to expire at the end of the current calendar month.

$sel:httpStatus:GetDeliverabilityDashboardOptionsResponse', getDeliverabilityDashboardOptionsResponse_httpStatus - The response's http status code.

$sel:dashboardEnabled:GetDeliverabilityDashboardOptionsResponse', getDeliverabilityDashboardOptionsResponse_dashboardEnabled - Specifies whether the Deliverability dashboard is enabled. If this value is true, the dashboard is enabled.

Response Lenses

getDeliverabilityDashboardOptionsResponse_subscriptionExpiryDate :: Lens' GetDeliverabilityDashboardOptionsResponse (Maybe UTCTime)

The date, in Unix time format, when your current subscription to the Deliverability dashboard is scheduled to expire, if your subscription is scheduled to expire at the end of the current calendar month. This value is null if you have an active subscription that isn’t due to expire at the end of the month.

getDeliverabilityDashboardOptionsResponse_accountStatus :: Lens' GetDeliverabilityDashboardOptionsResponse (Maybe DeliverabilityDashboardAccountStatus)

The current status of your Deliverability dashboard subscription. If this value is PENDING_EXPIRATION, your subscription is scheduled to expire at the end of the current calendar month.

getDeliverabilityDashboardOptionsResponse_activeSubscribedDomains :: Lens' GetDeliverabilityDashboardOptionsResponse (Maybe [DomainDeliverabilityTrackingOption])

An array of objects, one for each verified domain that you use to send email and currently has an active Deliverability dashboard subscription that isn’t scheduled to expire at the end of the current calendar month.

getDeliverabilityDashboardOptionsResponse_pendingExpirationSubscribedDomains :: Lens' GetDeliverabilityDashboardOptionsResponse (Maybe [DomainDeliverabilityTrackingOption])

An array of objects, one for each verified domain that you use to send email and currently has an active Deliverability dashboard subscription that's scheduled to expire at the end of the current calendar month.

getDeliverabilityDashboardOptionsResponse_httpStatus :: Lens' GetDeliverabilityDashboardOptionsResponse Int

The response's http status code.

getDeliverabilityDashboardOptionsResponse_dashboardEnabled :: Lens' GetDeliverabilityDashboardOptionsResponse Bool

Specifies whether the Deliverability dashboard is enabled. If this value is true, the dashboard is enabled.