Network.AWS.SESv2.GetDeliverabilityTestReport

Description

Retrieve the results of a predictive inbox placement test.

Creating a Request

data GetDeliverabilityTestReport

A request to retrieve the results of a predictive inbox placement test.

See: newGetDeliverabilityTestReport smart constructor.

Constructors

GetDeliverabilityTestReport'

Fields

reportId :: Text

A unique string that identifies the predictive inbox placement test.

newGetDeliverabilityTestReport

Arguments

:: Text
$sel:reportId:GetDeliverabilityTestReport'

-> GetDeliverabilityTestReport

Create a value of GetDeliverabilityTestReport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportId:GetDeliverabilityTestReport', getDeliverabilityTestReport_reportId - A unique string that identifies the predictive inbox placement test.

Request Lenses

getDeliverabilityTestReport_reportId :: Lens' GetDeliverabilityTestReport Text

A unique string that identifies the predictive inbox placement test.

Destructuring the Response

data GetDeliverabilityTestReportResponse

The results of the predictive inbox placement test.

See: newGetDeliverabilityTestReportResponse smart constructor.

Constructors

GetDeliverabilityTestReportResponse'

Fields

message :: Maybe Text

An object that contains the message that you sent when you performed this predictive inbox placement test.

tags :: Maybe [Tag]

An array of objects that define the tags (keys and values) that are associated with the predictive inbox placement test.

httpStatus :: Int

The response's http status code.

deliverabilityTestReport :: DeliverabilityTestReport

An object that contains the results of the predictive inbox placement test.

overallPlacement :: PlacementStatistics

An object that specifies how many test messages that were sent during the predictive inbox placement test were delivered to recipients' inboxes, how many were sent to recipients' spam folders, and how many weren't delivered.

ispPlacements :: [IspPlacement]

An object that describes how the test email was handled by several email providers, including Gmail, Hotmail, Yahoo, AOL, and others.

newGetDeliverabilityTestReportResponse

Arguments

:: Int
$sel:httpStatus:GetDeliverabilityTestReportResponse'

-> DeliverabilityTestReport
$sel:deliverabilityTestReport:GetDeliverabilityTestReportResponse'

-> PlacementStatistics
$sel:overallPlacement:GetDeliverabilityTestReportResponse'

-> GetDeliverabilityTestReportResponse

Create a value of GetDeliverabilityTestReportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:GetDeliverabilityTestReportResponse', getDeliverabilityTestReportResponse_message - An object that contains the message that you sent when you performed this predictive inbox placement test.

$sel:tags:GetDeliverabilityTestReportResponse', getDeliverabilityTestReportResponse_tags - An array of objects that define the tags (keys and values) that are associated with the predictive inbox placement test.

$sel:httpStatus:GetDeliverabilityTestReportResponse', getDeliverabilityTestReportResponse_httpStatus - The response's http status code.

$sel:deliverabilityTestReport:GetDeliverabilityTestReportResponse', getDeliverabilityTestReportResponse_deliverabilityTestReport - An object that contains the results of the predictive inbox placement test.

$sel:overallPlacement:GetDeliverabilityTestReportResponse', getDeliverabilityTestReportResponse_overallPlacement - An object that specifies how many test messages that were sent during the predictive inbox placement test were delivered to recipients' inboxes, how many were sent to recipients' spam folders, and how many weren't delivered.

$sel:ispPlacements:GetDeliverabilityTestReportResponse', getDeliverabilityTestReportResponse_ispPlacements - An object that describes how the test email was handled by several email providers, including Gmail, Hotmail, Yahoo, AOL, and others.

Response Lenses

getDeliverabilityTestReportResponse_message :: Lens' GetDeliverabilityTestReportResponse (Maybe Text)

An object that contains the message that you sent when you performed this predictive inbox placement test.

getDeliverabilityTestReportResponse_tags :: Lens' GetDeliverabilityTestReportResponse (Maybe [Tag])

An array of objects that define the tags (keys and values) that are associated with the predictive inbox placement test.

getDeliverabilityTestReportResponse_httpStatus :: Lens' GetDeliverabilityTestReportResponse Int

The response's http status code.

getDeliverabilityTestReportResponse_deliverabilityTestReport :: Lens' GetDeliverabilityTestReportResponse DeliverabilityTestReport

An object that contains the results of the predictive inbox placement test.

getDeliverabilityTestReportResponse_overallPlacement :: Lens' GetDeliverabilityTestReportResponse PlacementStatistics

An object that specifies how many test messages that were sent during the predictive inbox placement test were delivered to recipients' inboxes, how many were sent to recipients' spam folders, and how many weren't delivered.

getDeliverabilityTestReportResponse_ispPlacements :: Lens' GetDeliverabilityTestReportResponse [IspPlacement]

An object that describes how the test email was handled by several email providers, including Gmail, Hotmail, Yahoo, AOL, and others.