Network.AWS.SESv2.GetEmailIdentity

Description

Provides information about a specific identity, including the identity's verification status, sending authorization policies, its DKIM authentication status, and its custom Mail-From settings.

Creating a Request

data GetEmailIdentity

A request to return details about an email identity.

See: newGetEmailIdentity smart constructor.

Constructors

GetEmailIdentity'

Fields

emailIdentity :: Text

The email identity that you want to retrieve details for.

newGetEmailIdentity

Arguments

:: Text
$sel:emailIdentity:GetEmailIdentity'

-> GetEmailIdentity

Create a value of GetEmailIdentity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:emailIdentity:GetEmailIdentity', getEmailIdentity_emailIdentity - The email identity that you want to retrieve details for.

Request Lenses

getEmailIdentity_emailIdentity :: Lens' GetEmailIdentity Text

The email identity that you want to retrieve details for.

Destructuring the Response

data GetEmailIdentityResponse

Details about an email identity.

See: newGetEmailIdentityResponse smart constructor.

Constructors

GetEmailIdentityResponse'

Fields

dkimAttributes :: Maybe DkimAttributes

An object that contains information about the DKIM attributes for the identity.

feedbackForwardingStatus :: Maybe Bool

The feedback forwarding configuration for the identity.

If the value is true, you receive email notifications when bounce or complaint events occur. These notifications are sent to the address that you specified in the Return-Path header of the original email.

You're required to have a method of tracking bounces and complaints. If you haven't set up another mechanism for receiving bounce or complaint notifications (for example, by setting up an event destination), you receive an email notification when these events occur (even if this setting is disabled).

policies :: Maybe (HashMap Text Text)

A map of policy names to policies.

tags :: Maybe [Tag]

An array of objects that define the tags (keys and values) that are associated with the email identity.

identityType :: Maybe IdentityType

The email identity type.

mailFromAttributes :: Maybe MailFromAttributes

An object that contains information about the Mail-From attributes for the email identity.

configurationSetName :: Maybe Text

The configuration set used by default when sending from this identity.

verifiedForSendingStatus :: Maybe Bool

Specifies whether or not the identity is verified. You can only send email from verified email addresses or domains. For more information about verifying identities, see the Amazon Pinpoint User Guide.

httpStatus :: Int

The response's http status code.

newGetEmailIdentityResponse

Arguments

:: Int
$sel:httpStatus:GetEmailIdentityResponse'

-> GetEmailIdentityResponse

Create a value of GetEmailIdentityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dkimAttributes:GetEmailIdentityResponse', getEmailIdentityResponse_dkimAttributes - An object that contains information about the DKIM attributes for the identity.

$sel:feedbackForwardingStatus:GetEmailIdentityResponse', getEmailIdentityResponse_feedbackForwardingStatus - The feedback forwarding configuration for the identity.

If the value is true, you receive email notifications when bounce or complaint events occur. These notifications are sent to the address that you specified in the Return-Path header of the original email.

You're required to have a method of tracking bounces and complaints. If you haven't set up another mechanism for receiving bounce or complaint notifications (for example, by setting up an event destination), you receive an email notification when these events occur (even if this setting is disabled).

$sel:policies:GetEmailIdentityResponse', getEmailIdentityResponse_policies - A map of policy names to policies.

$sel:tags:GetEmailIdentityResponse', getEmailIdentityResponse_tags - An array of objects that define the tags (keys and values) that are associated with the email identity.

$sel:identityType:GetEmailIdentityResponse', getEmailIdentityResponse_identityType - The email identity type.

$sel:mailFromAttributes:GetEmailIdentityResponse', getEmailIdentityResponse_mailFromAttributes - An object that contains information about the Mail-From attributes for the email identity.

$sel:configurationSetName:GetEmailIdentityResponse', getEmailIdentityResponse_configurationSetName - The configuration set used by default when sending from this identity.

$sel:verifiedForSendingStatus:GetEmailIdentityResponse', getEmailIdentityResponse_verifiedForSendingStatus - Specifies whether or not the identity is verified. You can only send email from verified email addresses or domains. For more information about verifying identities, see the Amazon Pinpoint User Guide.

$sel:httpStatus:GetEmailIdentityResponse', getEmailIdentityResponse_httpStatus - The response's http status code.

Response Lenses

getEmailIdentityResponse_dkimAttributes :: Lens' GetEmailIdentityResponse (Maybe DkimAttributes)

An object that contains information about the DKIM attributes for the identity.

getEmailIdentityResponse_feedbackForwardingStatus :: Lens' GetEmailIdentityResponse (Maybe Bool)

The feedback forwarding configuration for the identity.

If the value is true, you receive email notifications when bounce or complaint events occur. These notifications are sent to the address that you specified in the Return-Path header of the original email.

You're required to have a method of tracking bounces and complaints. If you haven't set up another mechanism for receiving bounce or complaint notifications (for example, by setting up an event destination), you receive an email notification when these events occur (even if this setting is disabled).

getEmailIdentityResponse_policies :: Lens' GetEmailIdentityResponse (Maybe (HashMap Text Text))

A map of policy names to policies.

getEmailIdentityResponse_tags :: Lens' GetEmailIdentityResponse (Maybe [Tag])

An array of objects that define the tags (keys and values) that are associated with the email identity.

getEmailIdentityResponse_identityType :: Lens' GetEmailIdentityResponse (Maybe IdentityType)

The email identity type.

getEmailIdentityResponse_mailFromAttributes :: Lens' GetEmailIdentityResponse (Maybe MailFromAttributes)

An object that contains information about the Mail-From attributes for the email identity.

getEmailIdentityResponse_configurationSetName :: Lens' GetEmailIdentityResponse (Maybe Text)

The configuration set used by default when sending from this identity.

getEmailIdentityResponse_verifiedForSendingStatus :: Lens' GetEmailIdentityResponse (Maybe Bool)

Specifies whether or not the identity is verified. You can only send email from verified email addresses or domains. For more information about verifying identities, see the Amazon Pinpoint User Guide.

getEmailIdentityResponse_httpStatus :: Lens' GetEmailIdentityResponse Int

The response's http status code.