Network.AWS.SESv2.GetImportJob

Description

Provides information about an import job.

Creating a Request

data GetImportJob

Represents a request for information about an import job using the import job ID.

See: newGetImportJob smart constructor.

Constructors

GetImportJob'

Fields

jobId :: Text

The ID of the import job.

newGetImportJob

Arguments

:: Text
$sel:jobId:GetImportJob'

-> GetImportJob

Create a value of GetImportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:GetImportJob', getImportJob_jobId - The ID of the import job.

Request Lenses

getImportJob_jobId :: Lens' GetImportJob Text

The ID of the import job.

Destructuring the Response

data GetImportJobResponse

An HTTP 200 response if the request succeeds, or an error message if the request fails.

See: newGetImportJobResponse smart constructor.

Constructors

GetImportJobResponse'

Fields

failedRecordsCount :: Maybe Int

The number of records that failed processing because of invalid input or other reasons.

failureInfo :: Maybe FailureInfo

The failure details about an import job.

createdTimestamp :: Maybe POSIX

The time stamp of when the import job was created.

jobStatus :: Maybe JobStatus

The status of the import job.

completedTimestamp :: Maybe POSIX

The time stamp of when the import job was completed.

importDataSource :: Maybe ImportDataSource

The data source of the import job.

processedRecordsCount :: Maybe Int

The current number of records processed.

importDestination :: Maybe ImportDestination

The destination of the import job.

jobId :: Maybe Text

A string that represents the import job ID.

httpStatus :: Int

The response's http status code.

newGetImportJobResponse

Arguments

:: Int
$sel:httpStatus:GetImportJobResponse'

-> GetImportJobResponse

Create a value of GetImportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedRecordsCount:GetImportJobResponse', getImportJobResponse_failedRecordsCount - The number of records that failed processing because of invalid input or other reasons.

$sel:failureInfo:GetImportJobResponse', getImportJobResponse_failureInfo - The failure details about an import job.

$sel:createdTimestamp:GetImportJobResponse', getImportJobResponse_createdTimestamp - The time stamp of when the import job was created.

$sel:jobStatus:GetImportJobResponse', getImportJobResponse_jobStatus - The status of the import job.

$sel:completedTimestamp:GetImportJobResponse', getImportJobResponse_completedTimestamp - The time stamp of when the import job was completed.

$sel:importDataSource:GetImportJobResponse', getImportJobResponse_importDataSource - The data source of the import job.

$sel:processedRecordsCount:GetImportJobResponse', getImportJobResponse_processedRecordsCount - The current number of records processed.

$sel:importDestination:GetImportJobResponse', getImportJobResponse_importDestination - The destination of the import job.

$sel:jobId:GetImportJob', getImportJobResponse_jobId - A string that represents the import job ID.

$sel:httpStatus:GetImportJobResponse', getImportJobResponse_httpStatus - The response's http status code.

Response Lenses

getImportJobResponse_failedRecordsCount :: Lens' GetImportJobResponse (Maybe Int)

The number of records that failed processing because of invalid input or other reasons.

getImportJobResponse_failureInfo :: Lens' GetImportJobResponse (Maybe FailureInfo)

The failure details about an import job.

getImportJobResponse_createdTimestamp :: Lens' GetImportJobResponse (Maybe UTCTime)

The time stamp of when the import job was created.

getImportJobResponse_jobStatus :: Lens' GetImportJobResponse (Maybe JobStatus)

The status of the import job.

getImportJobResponse_completedTimestamp :: Lens' GetImportJobResponse (Maybe UTCTime)

The time stamp of when the import job was completed.

getImportJobResponse_importDataSource :: Lens' GetImportJobResponse (Maybe ImportDataSource)

The data source of the import job.

getImportJobResponse_processedRecordsCount :: Lens' GetImportJobResponse (Maybe Int)

The current number of records processed.

getImportJobResponse_importDestination :: Lens' GetImportJobResponse (Maybe ImportDestination)

The destination of the import job.

getImportJobResponse_jobId :: Lens' GetImportJobResponse (Maybe Text)

A string that represents the import job ID.

getImportJobResponse_httpStatus :: Lens' GetImportJobResponse Int

The response's http status code.