Network.AWS.SESv2.GetSuppressedDestination

Description

Retrieves information about a specific email address that's on the suppression list for your account.

Creating a Request

data GetSuppressedDestination

A request to retrieve information about an email address that's on the suppression list for your account.

See: newGetSuppressedDestination smart constructor.

Constructors

GetSuppressedDestination'

Fields

emailAddress :: Text

The email address that's on the account suppression list.

newGetSuppressedDestination

Arguments

:: Text
$sel:emailAddress:GetSuppressedDestination'

-> GetSuppressedDestination

Create a value of GetSuppressedDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:emailAddress:GetSuppressedDestination', getSuppressedDestination_emailAddress - The email address that's on the account suppression list.

Request Lenses

getSuppressedDestination_emailAddress :: Lens' GetSuppressedDestination Text

The email address that's on the account suppression list.

Destructuring the Response

data GetSuppressedDestinationResponse

Information about the suppressed email address.

See: newGetSuppressedDestinationResponse smart constructor.

Constructors

GetSuppressedDestinationResponse'

Fields

httpStatus :: Int

The response's http status code.

suppressedDestination :: SuppressedDestination

An object containing information about the suppressed email address.

newGetSuppressedDestinationResponse

Arguments

:: Int
$sel:httpStatus:GetSuppressedDestinationResponse'

-> SuppressedDestination
$sel:suppressedDestination:GetSuppressedDestinationResponse'

-> GetSuppressedDestinationResponse

Create a value of GetSuppressedDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetSuppressedDestinationResponse', getSuppressedDestinationResponse_httpStatus - The response's http status code.

$sel:suppressedDestination:GetSuppressedDestinationResponse', getSuppressedDestinationResponse_suppressedDestination - An object containing information about the suppressed email address.

Response Lenses

getSuppressedDestinationResponse_httpStatus :: Lens' GetSuppressedDestinationResponse Int

The response's http status code.

getSuppressedDestinationResponse_suppressedDestination :: Lens' GetSuppressedDestinationResponse SuppressedDestination

An object containing information about the suppressed email address.