Network.AWS.SESv2.ListConfigurationSets

Description

List all of the configuration sets associated with your account in the current region.

Configuration sets are groups of rules that you can apply to the emails you send. You apply a configuration set to an email by including a reference to the configuration set in the headers of the email. When you apply a configuration set to an email, all of the rules in that configuration set are applied to the email.

Creating a Request

data ListConfigurationSets

A request to obtain a list of configuration sets for your Amazon SES account in the current AWS Region.

See: newListConfigurationSets smart constructor.

Constructors

ListConfigurationSets'

Fields

nextToken :: Maybe Text

A token returned from a previous call to ListConfigurationSets to indicate the position in the list of configuration sets.

pageSize :: Maybe Int

The number of results to show in a single call to ListConfigurationSets. If the number of results is larger than the number you specified in this parameter, then the response includes a NextToken element, which you can use to obtain additional results.

newListConfigurationSets :: ListConfigurationSets

Create a value of ListConfigurationSets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListConfigurationSets', listConfigurationSets_nextToken - A token returned from a previous call to ListConfigurationSets to indicate the position in the list of configuration sets.

$sel:pageSize:ListConfigurationSets', listConfigurationSets_pageSize - The number of results to show in a single call to ListConfigurationSets. If the number of results is larger than the number you specified in this parameter, then the response includes a NextToken element, which you can use to obtain additional results.

Request Lenses

listConfigurationSets_nextToken :: Lens' ListConfigurationSets (Maybe Text)

A token returned from a previous call to ListConfigurationSets to indicate the position in the list of configuration sets.

listConfigurationSets_pageSize :: Lens' ListConfigurationSets (Maybe Int)

The number of results to show in a single call to ListConfigurationSets. If the number of results is larger than the number you specified in this parameter, then the response includes a NextToken element, which you can use to obtain additional results.

Destructuring the Response

data ListConfigurationSetsResponse

A list of configuration sets in your Amazon SES account in the current AWS Region.

See: newListConfigurationSetsResponse smart constructor.

Constructors

ListConfigurationSetsResponse'

Fields

nextToken :: Maybe Text

A token that indicates that there are additional configuration sets to list. To view additional configuration sets, issue another request to ListConfigurationSets, and pass this token in the NextToken parameter.

configurationSets :: Maybe [Text]

An array that contains all of the configuration sets in your Amazon SES account in the current AWS Region.

httpStatus :: Int

The response's http status code.

newListConfigurationSetsResponse

Arguments

:: Int
$sel:httpStatus:ListConfigurationSetsResponse'

-> ListConfigurationSetsResponse

Create a value of ListConfigurationSetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListConfigurationSets', listConfigurationSetsResponse_nextToken - A token that indicates that there are additional configuration sets to list. To view additional configuration sets, issue another request to ListConfigurationSets, and pass this token in the NextToken parameter.

$sel:configurationSets:ListConfigurationSetsResponse', listConfigurationSetsResponse_configurationSets - An array that contains all of the configuration sets in your Amazon SES account in the current AWS Region.

$sel:httpStatus:ListConfigurationSetsResponse', listConfigurationSetsResponse_httpStatus - The response's http status code.

Response Lenses

listConfigurationSetsResponse_nextToken :: Lens' ListConfigurationSetsResponse (Maybe Text)

A token that indicates that there are additional configuration sets to list. To view additional configuration sets, issue another request to ListConfigurationSets, and pass this token in the NextToken parameter.

listConfigurationSetsResponse_configurationSets :: Lens' ListConfigurationSetsResponse (Maybe [Text])

An array that contains all of the configuration sets in your Amazon SES account in the current AWS Region.

listConfigurationSetsResponse_httpStatus :: Lens' ListConfigurationSetsResponse Int

The response's http status code.