Network.AWS.SESv2.ListContactLists

Description

Lists all of the contact lists available.

Creating a Request

data ListContactLists

See: newListContactLists smart constructor.

Constructors

ListContactLists'

Fields

nextToken :: Maybe Text

A string token indicating that there might be additional contact lists available to be listed. Use the token provided in the Response to use in the subsequent call to ListContactLists with the same parameters to retrieve the next page of contact lists.

pageSize :: Maybe Int

Maximum number of contact lists to return at once. Use this parameter to paginate results. If additional contact lists exist beyond the specified limit, the NextToken element is sent in the response. Use the NextToken value in subsequent requests to retrieve additional lists.

newListContactLists :: ListContactLists

Create a value of ListContactLists with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListContactLists', listContactLists_nextToken - A string token indicating that there might be additional contact lists available to be listed. Use the token provided in the Response to use in the subsequent call to ListContactLists with the same parameters to retrieve the next page of contact lists.

$sel:pageSize:ListContactLists', listContactLists_pageSize - Maximum number of contact lists to return at once. Use this parameter to paginate results. If additional contact lists exist beyond the specified limit, the NextToken element is sent in the response. Use the NextToken value in subsequent requests to retrieve additional lists.

Request Lenses

listContactLists_nextToken :: Lens' ListContactLists (Maybe Text)

A string token indicating that there might be additional contact lists available to be listed. Use the token provided in the Response to use in the subsequent call to ListContactLists with the same parameters to retrieve the next page of contact lists.

listContactLists_pageSize :: Lens' ListContactLists (Maybe Int)

Maximum number of contact lists to return at once. Use this parameter to paginate results. If additional contact lists exist beyond the specified limit, the NextToken element is sent in the response. Use the NextToken value in subsequent requests to retrieve additional lists.

Destructuring the Response

data ListContactListsResponse

See: newListContactListsResponse smart constructor.

Constructors

ListContactListsResponse'

Fields

nextToken :: Maybe Text

A string token indicating that there might be additional contact lists available to be listed. Copy this token to a subsequent call to ListContactLists with the same parameters to retrieve the next page of contact lists.

contactLists :: Maybe [ContactList]

The available contact lists.

httpStatus :: Int

The response's http status code.

newListContactListsResponse

Arguments

:: Int
$sel:httpStatus:ListContactListsResponse'

-> ListContactListsResponse

Create a value of ListContactListsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListContactLists', listContactListsResponse_nextToken - A string token indicating that there might be additional contact lists available to be listed. Copy this token to a subsequent call to ListContactLists with the same parameters to retrieve the next page of contact lists.

$sel:contactLists:ListContactListsResponse', listContactListsResponse_contactLists - The available contact lists.

$sel:httpStatus:ListContactListsResponse', listContactListsResponse_httpStatus - The response's http status code.

Response Lenses

listContactListsResponse_nextToken :: Lens' ListContactListsResponse (Maybe Text)

A string token indicating that there might be additional contact lists available to be listed. Copy this token to a subsequent call to ListContactLists with the same parameters to retrieve the next page of contact lists.

listContactListsResponse_contactLists :: Lens' ListContactListsResponse (Maybe [ContactList])

The available contact lists.

listContactListsResponse_httpStatus :: Lens' ListContactListsResponse Int

The response's http status code.