Network.AWS.SESv2.ListDedicatedIpPools

Description

List all of the dedicated IP pools that exist in your Amazon Web Services account in the current Region.

Creating a Request

data ListDedicatedIpPools

A request to obtain a list of dedicated IP pools.

See: newListDedicatedIpPools smart constructor.

Constructors

ListDedicatedIpPools'

Fields

nextToken :: Maybe Text

A token returned from a previous call to ListDedicatedIpPools to indicate the position in the list of dedicated IP pools.

pageSize :: Maybe Int

The number of results to show in a single call to ListDedicatedIpPools. If the number of results is larger than the number you specified in this parameter, then the response includes a NextToken element, which you can use to obtain additional results.

newListDedicatedIpPools :: ListDedicatedIpPools

Create a value of ListDedicatedIpPools with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDedicatedIpPools', listDedicatedIpPools_nextToken - A token returned from a previous call to ListDedicatedIpPools to indicate the position in the list of dedicated IP pools.

$sel:pageSize:ListDedicatedIpPools', listDedicatedIpPools_pageSize - The number of results to show in a single call to ListDedicatedIpPools. If the number of results is larger than the number you specified in this parameter, then the response includes a NextToken element, which you can use to obtain additional results.

Request Lenses

listDedicatedIpPools_nextToken :: Lens' ListDedicatedIpPools (Maybe Text)

A token returned from a previous call to ListDedicatedIpPools to indicate the position in the list of dedicated IP pools.

listDedicatedIpPools_pageSize :: Lens' ListDedicatedIpPools (Maybe Int)

The number of results to show in a single call to ListDedicatedIpPools. If the number of results is larger than the number you specified in this parameter, then the response includes a NextToken element, which you can use to obtain additional results.

Destructuring the Response

data ListDedicatedIpPoolsResponse

A list of dedicated IP pools.

See: newListDedicatedIpPoolsResponse smart constructor.

Constructors

ListDedicatedIpPoolsResponse'

Fields

dedicatedIpPools :: Maybe [Text]

A list of all of the dedicated IP pools that are associated with your Amazon Web Services account in the current Region.

nextToken :: Maybe Text

A token that indicates that there are additional IP pools to list. To view additional IP pools, issue another request to ListDedicatedIpPools, passing this token in the NextToken parameter.

httpStatus :: Int

The response's http status code.

newListDedicatedIpPoolsResponse

Arguments

:: Int
$sel:httpStatus:ListDedicatedIpPoolsResponse'

-> ListDedicatedIpPoolsResponse

Create a value of ListDedicatedIpPoolsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dedicatedIpPools:ListDedicatedIpPoolsResponse', listDedicatedIpPoolsResponse_dedicatedIpPools - A list of all of the dedicated IP pools that are associated with your Amazon Web Services account in the current Region.

$sel:nextToken:ListDedicatedIpPools', listDedicatedIpPoolsResponse_nextToken - A token that indicates that there are additional IP pools to list. To view additional IP pools, issue another request to ListDedicatedIpPools, passing this token in the NextToken parameter.

$sel:httpStatus:ListDedicatedIpPoolsResponse', listDedicatedIpPoolsResponse_httpStatus - The response's http status code.

Response Lenses

listDedicatedIpPoolsResponse_dedicatedIpPools :: Lens' ListDedicatedIpPoolsResponse (Maybe [Text])

A list of all of the dedicated IP pools that are associated with your Amazon Web Services account in the current Region.

listDedicatedIpPoolsResponse_nextToken :: Lens' ListDedicatedIpPoolsResponse (Maybe Text)

A token that indicates that there are additional IP pools to list. To view additional IP pools, issue another request to ListDedicatedIpPools, passing this token in the NextToken parameter.

listDedicatedIpPoolsResponse_httpStatus :: Lens' ListDedicatedIpPoolsResponse Int

The response's http status code.