Network.AWS.SESv2.ListDomainDeliverabilityCampaigns

Description

Retrieve deliverability data for all the campaigns that used a specific domain to send email during a specified time range. This data is available for a domain only if you enabled the Deliverability dashboard for the domain.

Creating a Request

data ListDomainDeliverabilityCampaigns

Retrieve deliverability data for all the campaigns that used a specific domain to send email during a specified time range. This data is available for a domain only if you enabled the Deliverability dashboard.

See: newListDomainDeliverabilityCampaigns smart constructor.

Constructors

ListDomainDeliverabilityCampaigns'

Fields

nextToken :: Maybe Text

A token that’s returned from a previous call to the ListDomainDeliverabilityCampaigns operation. This token indicates the position of a campaign in the list of campaigns.

pageSize :: Maybe Int

The maximum number of results to include in response to a single call to the ListDomainDeliverabilityCampaigns operation. If the number of results is larger than the number that you specify in this parameter, the response includes a NextToken element, which you can use to obtain additional results.

startDate :: POSIX

The first day, in Unix time format, that you want to obtain deliverability data for.

endDate :: POSIX

The last day, in Unix time format, that you want to obtain deliverability data for. This value has to be less than or equal to 30 days after the value of the StartDate parameter.

subscribedDomain :: Text

The domain to obtain deliverability data for.

newListDomainDeliverabilityCampaigns

Arguments

:: UTCTime
$sel:startDate:ListDomainDeliverabilityCampaigns'

-> UTCTime
$sel:endDate:ListDomainDeliverabilityCampaigns'

-> Text
$sel:subscribedDomain:ListDomainDeliverabilityCampaigns'

-> ListDomainDeliverabilityCampaigns

Create a value of ListDomainDeliverabilityCampaigns with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDomainDeliverabilityCampaigns', listDomainDeliverabilityCampaigns_nextToken - A token that’s returned from a previous call to the ListDomainDeliverabilityCampaigns operation. This token indicates the position of a campaign in the list of campaigns.

$sel:pageSize:ListDomainDeliverabilityCampaigns', listDomainDeliverabilityCampaigns_pageSize - The maximum number of results to include in response to a single call to the ListDomainDeliverabilityCampaigns operation. If the number of results is larger than the number that you specify in this parameter, the response includes a NextToken element, which you can use to obtain additional results.

$sel:startDate:ListDomainDeliverabilityCampaigns', listDomainDeliverabilityCampaigns_startDate - The first day, in Unix time format, that you want to obtain deliverability data for.

$sel:endDate:ListDomainDeliverabilityCampaigns', listDomainDeliverabilityCampaigns_endDate - The last day, in Unix time format, that you want to obtain deliverability data for. This value has to be less than or equal to 30 days after the value of the StartDate parameter.

$sel:subscribedDomain:ListDomainDeliverabilityCampaigns', listDomainDeliverabilityCampaigns_subscribedDomain - The domain to obtain deliverability data for.

Request Lenses

listDomainDeliverabilityCampaigns_nextToken :: Lens' ListDomainDeliverabilityCampaigns (Maybe Text)

A token that’s returned from a previous call to the ListDomainDeliverabilityCampaigns operation. This token indicates the position of a campaign in the list of campaigns.

listDomainDeliverabilityCampaigns_pageSize :: Lens' ListDomainDeliverabilityCampaigns (Maybe Int)

The maximum number of results to include in response to a single call to the ListDomainDeliverabilityCampaigns operation. If the number of results is larger than the number that you specify in this parameter, the response includes a NextToken element, which you can use to obtain additional results.

listDomainDeliverabilityCampaigns_startDate :: Lens' ListDomainDeliverabilityCampaigns UTCTime

The first day, in Unix time format, that you want to obtain deliverability data for.

listDomainDeliverabilityCampaigns_endDate :: Lens' ListDomainDeliverabilityCampaigns UTCTime

The last day, in Unix time format, that you want to obtain deliverability data for. This value has to be less than or equal to 30 days after the value of the StartDate parameter.

listDomainDeliverabilityCampaigns_subscribedDomain :: Lens' ListDomainDeliverabilityCampaigns Text

The domain to obtain deliverability data for.

Destructuring the Response

data ListDomainDeliverabilityCampaignsResponse

An array of objects that provide deliverability data for all the campaigns that used a specific domain to send email during a specified time range. This data is available for a domain only if you enabled the Deliverability dashboard for the domain.

See: newListDomainDeliverabilityCampaignsResponse smart constructor.

Constructors

ListDomainDeliverabilityCampaignsResponse'

Fields

nextToken :: Maybe Text

A token that’s returned from a previous call to the ListDomainDeliverabilityCampaigns operation. This token indicates the position of the campaign in the list of campaigns.

httpStatus :: Int

The response's http status code.

domainDeliverabilityCampaigns :: [DomainDeliverabilityCampaign]

An array of responses, one for each campaign that used the domain to send email during the specified time range.

newListDomainDeliverabilityCampaignsResponse

Arguments

:: Int
$sel:httpStatus:ListDomainDeliverabilityCampaignsResponse'

-> ListDomainDeliverabilityCampaignsResponse

Create a value of ListDomainDeliverabilityCampaignsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDomainDeliverabilityCampaigns', listDomainDeliverabilityCampaignsResponse_nextToken - A token that’s returned from a previous call to the ListDomainDeliverabilityCampaigns operation. This token indicates the position of the campaign in the list of campaigns.

$sel:httpStatus:ListDomainDeliverabilityCampaignsResponse', listDomainDeliverabilityCampaignsResponse_httpStatus - The response's http status code.

$sel:domainDeliverabilityCampaigns:ListDomainDeliverabilityCampaignsResponse', listDomainDeliverabilityCampaignsResponse_domainDeliverabilityCampaigns - An array of responses, one for each campaign that used the domain to send email during the specified time range.

Response Lenses

listDomainDeliverabilityCampaignsResponse_nextToken :: Lens' ListDomainDeliverabilityCampaignsResponse (Maybe Text)

A token that’s returned from a previous call to the ListDomainDeliverabilityCampaigns operation. This token indicates the position of the campaign in the list of campaigns.

listDomainDeliverabilityCampaignsResponse_httpStatus :: Lens' ListDomainDeliverabilityCampaignsResponse Int

The response's http status code.

listDomainDeliverabilityCampaignsResponse_domainDeliverabilityCampaigns :: Lens' ListDomainDeliverabilityCampaignsResponse [DomainDeliverabilityCampaign]

An array of responses, one for each campaign that used the domain to send email during the specified time range.