Network.AWS.SESv2.ListImportJobs

Description

Lists all of the import jobs.

Creating a Request

data ListImportJobs

Represents a request to list all of the import jobs for a data destination within the specified maximum number of import jobs.

See: newListImportJobs smart constructor.

Constructors

ListImportJobs'

Fields

nextToken :: Maybe Text

A string token indicating that there might be additional import jobs available to be listed. Copy this token to a subsequent call to ListImportJobs with the same parameters to retrieve the next page of import jobs.

pageSize :: Maybe Int

Maximum number of import jobs to return at once. Use this parameter to paginate results. If additional import jobs exist beyond the specified limit, the NextToken element is sent in the response. Use the NextToken value in subsequent requests to retrieve additional addresses.

importDestinationType :: Maybe ImportDestinationType

The destination of the import job, which can be used to list import jobs that have a certain ImportDestinationType.

newListImportJobs :: ListImportJobs

Create a value of ListImportJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListImportJobs', listImportJobs_nextToken - A string token indicating that there might be additional import jobs available to be listed. Copy this token to a subsequent call to ListImportJobs with the same parameters to retrieve the next page of import jobs.

$sel:pageSize:ListImportJobs', listImportJobs_pageSize - Maximum number of import jobs to return at once. Use this parameter to paginate results. If additional import jobs exist beyond the specified limit, the NextToken element is sent in the response. Use the NextToken value in subsequent requests to retrieve additional addresses.

$sel:importDestinationType:ListImportJobs', listImportJobs_importDestinationType - The destination of the import job, which can be used to list import jobs that have a certain ImportDestinationType.

Request Lenses

listImportJobs_nextToken :: Lens' ListImportJobs (Maybe Text)

A string token indicating that there might be additional import jobs available to be listed. Copy this token to a subsequent call to ListImportJobs with the same parameters to retrieve the next page of import jobs.

listImportJobs_pageSize :: Lens' ListImportJobs (Maybe Int)

Maximum number of import jobs to return at once. Use this parameter to paginate results. If additional import jobs exist beyond the specified limit, the NextToken element is sent in the response. Use the NextToken value in subsequent requests to retrieve additional addresses.

listImportJobs_importDestinationType :: Lens' ListImportJobs (Maybe ImportDestinationType)

The destination of the import job, which can be used to list import jobs that have a certain ImportDestinationType.

Destructuring the Response

data ListImportJobsResponse

An HTTP 200 response if the request succeeds, or an error message if the request fails.

See: newListImportJobsResponse smart constructor.

Constructors

ListImportJobsResponse'

Fields

importJobs :: Maybe [ImportJobSummary]

A list of the import job summaries.

nextToken :: Maybe Text

A string token indicating that there might be additional import jobs available to be listed. Copy this token to a subsequent call to ListImportJobs with the same parameters to retrieve the next page of import jobs.

httpStatus :: Int

The response's http status code.

newListImportJobsResponse

Arguments

:: Int
$sel:httpStatus:ListImportJobsResponse'

-> ListImportJobsResponse

Create a value of ListImportJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:importJobs:ListImportJobsResponse', listImportJobsResponse_importJobs - A list of the import job summaries.

$sel:nextToken:ListImportJobs', listImportJobsResponse_nextToken - A string token indicating that there might be additional import jobs available to be listed. Copy this token to a subsequent call to ListImportJobs with the same parameters to retrieve the next page of import jobs.

$sel:httpStatus:ListImportJobsResponse', listImportJobsResponse_httpStatus - The response's http status code.

Response Lenses

listImportJobsResponse_importJobs :: Lens' ListImportJobsResponse (Maybe [ImportJobSummary])

A list of the import job summaries.

listImportJobsResponse_nextToken :: Lens' ListImportJobsResponse (Maybe Text)

A string token indicating that there might be additional import jobs available to be listed. Copy this token to a subsequent call to ListImportJobs with the same parameters to retrieve the next page of import jobs.

listImportJobsResponse_httpStatus :: Lens' ListImportJobsResponse Int

The response's http status code.