Network.AWS.SESv2.ListSuppressedDestinations

Description

Retrieves a list of email addresses that are on the suppression list for your account.

Creating a Request

data ListSuppressedDestinations

A request to obtain a list of email destinations that are on the suppression list for your account.

See: newListSuppressedDestinations smart constructor.

Constructors

ListSuppressedDestinations'

Fields

reasons :: Maybe [SuppressionListReason]

The factors that caused the email address to be added to .

endDate :: Maybe POSIX

Used to filter the list of suppressed email destinations so that it only includes addresses that were added to the list before a specific date. The date that you specify should be in Unix time format.

startDate :: Maybe POSIX

Used to filter the list of suppressed email destinations so that it only includes addresses that were added to the list after a specific date. The date that you specify should be in Unix time format.

nextToken :: Maybe Text

A token returned from a previous call to ListSuppressedDestinations to indicate the position in the list of suppressed email addresses.

pageSize :: Maybe Int

The number of results to show in a single call to ListSuppressedDestinations. If the number of results is larger than the number you specified in this parameter, then the response includes a NextToken element, which you can use to obtain additional results.

newListSuppressedDestinations :: ListSuppressedDestinations

Create a value of ListSuppressedDestinations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reasons:ListSuppressedDestinations', listSuppressedDestinations_reasons - The factors that caused the email address to be added to .

$sel:endDate:ListSuppressedDestinations', listSuppressedDestinations_endDate - Used to filter the list of suppressed email destinations so that it only includes addresses that were added to the list before a specific date. The date that you specify should be in Unix time format.

$sel:startDate:ListSuppressedDestinations', listSuppressedDestinations_startDate - Used to filter the list of suppressed email destinations so that it only includes addresses that were added to the list after a specific date. The date that you specify should be in Unix time format.

$sel:nextToken:ListSuppressedDestinations', listSuppressedDestinations_nextToken - A token returned from a previous call to ListSuppressedDestinations to indicate the position in the list of suppressed email addresses.

$sel:pageSize:ListSuppressedDestinations', listSuppressedDestinations_pageSize - The number of results to show in a single call to ListSuppressedDestinations. If the number of results is larger than the number you specified in this parameter, then the response includes a NextToken element, which you can use to obtain additional results.

Request Lenses

listSuppressedDestinations_reasons :: Lens' ListSuppressedDestinations (Maybe [SuppressionListReason])

The factors that caused the email address to be added to .

listSuppressedDestinations_endDate :: Lens' ListSuppressedDestinations (Maybe UTCTime)

Used to filter the list of suppressed email destinations so that it only includes addresses that were added to the list before a specific date. The date that you specify should be in Unix time format.

listSuppressedDestinations_startDate :: Lens' ListSuppressedDestinations (Maybe UTCTime)

Used to filter the list of suppressed email destinations so that it only includes addresses that were added to the list after a specific date. The date that you specify should be in Unix time format.

listSuppressedDestinations_nextToken :: Lens' ListSuppressedDestinations (Maybe Text)

A token returned from a previous call to ListSuppressedDestinations to indicate the position in the list of suppressed email addresses.

listSuppressedDestinations_pageSize :: Lens' ListSuppressedDestinations (Maybe Int)

The number of results to show in a single call to ListSuppressedDestinations. If the number of results is larger than the number you specified in this parameter, then the response includes a NextToken element, which you can use to obtain additional results.

Destructuring the Response

data ListSuppressedDestinationsResponse

A list of suppressed email addresses.

See: newListSuppressedDestinationsResponse smart constructor.

Constructors

ListSuppressedDestinationsResponse'

Fields

nextToken :: Maybe Text

A token that indicates that there are additional email addresses on the suppression list for your account. To view additional suppressed addresses, issue another request to ListSuppressedDestinations, and pass this token in the NextToken parameter.

suppressedDestinationSummaries :: Maybe [SuppressedDestinationSummary]

A list of summaries, each containing a summary for a suppressed email destination.

httpStatus :: Int

The response's http status code.

newListSuppressedDestinationsResponse

Arguments

:: Int
$sel:httpStatus:ListSuppressedDestinationsResponse'

-> ListSuppressedDestinationsResponse

Create a value of ListSuppressedDestinationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSuppressedDestinations', listSuppressedDestinationsResponse_nextToken - A token that indicates that there are additional email addresses on the suppression list for your account. To view additional suppressed addresses, issue another request to ListSuppressedDestinations, and pass this token in the NextToken parameter.

$sel:suppressedDestinationSummaries:ListSuppressedDestinationsResponse', listSuppressedDestinationsResponse_suppressedDestinationSummaries - A list of summaries, each containing a summary for a suppressed email destination.

$sel:httpStatus:ListSuppressedDestinationsResponse', listSuppressedDestinationsResponse_httpStatus - The response's http status code.

Response Lenses

listSuppressedDestinationsResponse_nextToken :: Lens' ListSuppressedDestinationsResponse (Maybe Text)

A token that indicates that there are additional email addresses on the suppression list for your account. To view additional suppressed addresses, issue another request to ListSuppressedDestinations, and pass this token in the NextToken parameter.

listSuppressedDestinationsResponse_suppressedDestinationSummaries :: Lens' ListSuppressedDestinationsResponse (Maybe [SuppressedDestinationSummary])

A list of summaries, each containing a summary for a suppressed email destination.

listSuppressedDestinationsResponse_httpStatus :: Lens' ListSuppressedDestinationsResponse Int

The response's http status code.