Network.AWS.SESv2.PutEmailIdentityFeedbackAttributes

Description

Used to enable or disable feedback forwarding for an identity. This setting determines what happens when an identity is used to send an email that results in a bounce or complaint event.

If the value is true, you receive email notifications when bounce or complaint events occur. These notifications are sent to the address that you specified in the Return-Path header of the original email.

You're required to have a method of tracking bounces and complaints. If you haven't set up another mechanism for receiving bounce or complaint notifications (for example, by setting up an event destination), you receive an email notification when these events occur (even if this setting is disabled).

Creating a Request

data PutEmailIdentityFeedbackAttributes

A request to set the attributes that control how bounce and complaint events are processed.

See: newPutEmailIdentityFeedbackAttributes smart constructor.

Constructors

PutEmailIdentityFeedbackAttributes'

Fields

emailForwardingEnabled :: Maybe Bool

Sets the feedback forwarding configuration for the identity.

If the value is true, you receive email notifications when bounce or complaint events occur. These notifications are sent to the address that you specified in the Return-Path header of the original email.

You're required to have a method of tracking bounces and complaints. If you haven't set up another mechanism for receiving bounce or complaint notifications (for example, by setting up an event destination), you receive an email notification when these events occur (even if this setting is disabled).

emailIdentity :: Text

The email identity that you want to configure bounce and complaint feedback forwarding for.

newPutEmailIdentityFeedbackAttributes

Arguments

:: Text
$sel:emailIdentity:PutEmailIdentityFeedbackAttributes'

-> PutEmailIdentityFeedbackAttributes

Create a value of PutEmailIdentityFeedbackAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:emailForwardingEnabled:PutEmailIdentityFeedbackAttributes', putEmailIdentityFeedbackAttributes_emailForwardingEnabled - Sets the feedback forwarding configuration for the identity.

If the value is true, you receive email notifications when bounce or complaint events occur. These notifications are sent to the address that you specified in the Return-Path header of the original email.

You're required to have a method of tracking bounces and complaints. If you haven't set up another mechanism for receiving bounce or complaint notifications (for example, by setting up an event destination), you receive an email notification when these events occur (even if this setting is disabled).

$sel:emailIdentity:PutEmailIdentityFeedbackAttributes', putEmailIdentityFeedbackAttributes_emailIdentity - The email identity that you want to configure bounce and complaint feedback forwarding for.

Request Lenses

putEmailIdentityFeedbackAttributes_emailForwardingEnabled :: Lens' PutEmailIdentityFeedbackAttributes (Maybe Bool)

Sets the feedback forwarding configuration for the identity.

If the value is true, you receive email notifications when bounce or complaint events occur. These notifications are sent to the address that you specified in the Return-Path header of the original email.

You're required to have a method of tracking bounces and complaints. If you haven't set up another mechanism for receiving bounce or complaint notifications (for example, by setting up an event destination), you receive an email notification when these events occur (even if this setting is disabled).

putEmailIdentityFeedbackAttributes_emailIdentity :: Lens' PutEmailIdentityFeedbackAttributes Text

The email identity that you want to configure bounce and complaint feedback forwarding for.

Destructuring the Response

data PutEmailIdentityFeedbackAttributesResponse

An HTTP 200 response if the request succeeds, or an error message if the request fails.

See: newPutEmailIdentityFeedbackAttributesResponse smart constructor.

Constructors

PutEmailIdentityFeedbackAttributesResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutEmailIdentityFeedbackAttributesResponse

Arguments

:: Int
$sel:httpStatus:PutEmailIdentityFeedbackAttributesResponse'

-> PutEmailIdentityFeedbackAttributesResponse

Create a value of PutEmailIdentityFeedbackAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutEmailIdentityFeedbackAttributesResponse', putEmailIdentityFeedbackAttributesResponse_httpStatus - The response's http status code.

Response Lenses

putEmailIdentityFeedbackAttributesResponse_httpStatus :: Lens' PutEmailIdentityFeedbackAttributesResponse Int

The response's http status code.