Network.AWS.SESv2.TestRenderEmailTemplate

Description

Creates a preview of the MIME content of an email when provided with a template and a set of replacement data.

You can execute this operation no more than once per second.

Creating a Request

data TestRenderEmailTemplate

Represents a request to create a preview of the MIME content of an

email when provided with a template and a set of replacement data.

See: newTestRenderEmailTemplate smart constructor.

Constructors

TestRenderEmailTemplate'

Fields

templateName :: Text

The name of the template that you want to render.

templateData :: Text

A list of replacement values to apply to the template. This parameter is a JSON object, typically consisting of key-value pairs in which the keys correspond to replacement tags in the email template.

newTestRenderEmailTemplate

Arguments

:: Text
$sel:templateName:TestRenderEmailTemplate'

-> Text
$sel:templateData:TestRenderEmailTemplate'

-> TestRenderEmailTemplate

Create a value of TestRenderEmailTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:TestRenderEmailTemplate', testRenderEmailTemplate_templateName - The name of the template that you want to render.

$sel:templateData:TestRenderEmailTemplate', testRenderEmailTemplate_templateData - A list of replacement values to apply to the template. This parameter is a JSON object, typically consisting of key-value pairs in which the keys correspond to replacement tags in the email template.

Request Lenses

testRenderEmailTemplate_templateName :: Lens' TestRenderEmailTemplate Text

The name of the template that you want to render.

testRenderEmailTemplate_templateData :: Lens' TestRenderEmailTemplate Text

A list of replacement values to apply to the template. This parameter is a JSON object, typically consisting of key-value pairs in which the keys correspond to replacement tags in the email template.

Destructuring the Response

data TestRenderEmailTemplateResponse

The following element is returned by the service.

See: newTestRenderEmailTemplateResponse smart constructor.

Constructors

TestRenderEmailTemplateResponse'

Fields

httpStatus :: Int

The response's http status code.

renderedTemplate :: Text

The complete MIME message rendered by applying the data in the TemplateData parameter to the template specified in the TemplateName parameter.

newTestRenderEmailTemplateResponse

Arguments

:: Int
$sel:httpStatus:TestRenderEmailTemplateResponse'

-> Text
$sel:renderedTemplate:TestRenderEmailTemplateResponse'

-> TestRenderEmailTemplateResponse

Create a value of TestRenderEmailTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TestRenderEmailTemplateResponse', testRenderEmailTemplateResponse_httpStatus - The response's http status code.

$sel:renderedTemplate:TestRenderEmailTemplateResponse', testRenderEmailTemplateResponse_renderedTemplate - The complete MIME message rendered by applying the data in the TemplateData parameter to the template specified in the TemplateName parameter.

Response Lenses

testRenderEmailTemplateResponse_httpStatus :: Lens' TestRenderEmailTemplateResponse Int

The response's http status code.

testRenderEmailTemplateResponse_renderedTemplate :: Lens' TestRenderEmailTemplateResponse Text

The complete MIME message rendered by applying the data in the TemplateData parameter to the template specified in the TemplateName parameter.