Network.AWS.SESv2.Types.Body

Description

Documentation

data Body

Represents the body of the email message.

See: newBody smart constructor.

Constructors

Body'

Fields

text :: Maybe Content

An object that represents the version of the message that is displayed in email clients that don't support HTML, or clients where the recipient has disabled HTML rendering.

html :: Maybe Content

An object that represents the version of the message that is displayed in email clients that support HTML. HTML messages can include formatted text, hyperlinks, images, and more.

newBody :: Body

Create a value of Body with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:text:Body', body_text - An object that represents the version of the message that is displayed in email clients that don't support HTML, or clients where the recipient has disabled HTML rendering.

$sel:html:Body', body_html - An object that represents the version of the message that is displayed in email clients that support HTML. HTML messages can include formatted text, hyperlinks, images, and more.

body_text :: Lens' Body (Maybe Content)

An object that represents the version of the message that is displayed in email clients that don't support HTML, or clients where the recipient has disabled HTML rendering.

body_html :: Lens' Body (Maybe Content)

An object that represents the version of the message that is displayed in email clients that support HTML. HTML messages can include formatted text, hyperlinks, images, and more.