Network.AWS.SESv2.Types.ContactListDestination

Description

Documentation

data ContactListDestination

An object that contains details about the action of a contact list.

See: newContactListDestination smart constructor.

Constructors

ContactListDestination'

Fields

contactListName :: Text

The name of the contact list.

contactListImportAction :: ContactListImportAction

The type of action to perform on the addresses. The following are the

possible values:

PUT: add the addresses to the contact list. If the record already exists, it will override it with the new value.

DELETE: remove the addresses from the contact list.

newContactListDestination

Arguments

:: Text
$sel:contactListName:ContactListDestination'

-> ContactListImportAction
$sel:contactListImportAction:ContactListDestination'

-> ContactListDestination

Create a value of ContactListDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contactListName:ContactListDestination', contactListDestination_contactListName - The name of the contact list.

$sel:contactListImportAction:ContactListDestination', contactListDestination_contactListImportAction - >The type of action to perform on the addresses. The following are the possible values:

PUT: add the addresses to the contact list. If the record already exists, it will override it with the new value.

DELETE: remove the addresses from the contact list.

contactListDestination_contactListName :: Lens' ContactListDestination Text

The name of the contact list.

contactListDestination_contactListImportAction :: Lens' ContactListDestination ContactListImportAction

The type of action to perform on the addresses. The following are the

possible values:

PUT: add the addresses to the contact list. If the record already exists, it will override it with the new value.

DELETE: remove the addresses from the contact list.