Network.AWS.SESv2.Types.DedicatedIp

Description

Documentation

data DedicatedIp

Contains information about a dedicated IP address that is associated with your Amazon SES account.

To learn more about requesting dedicated IP addresses, see Requesting and Relinquishing Dedicated IP Addresses in the Amazon SES Developer Guide.

See: newDedicatedIp smart constructor.

Constructors

DedicatedIp'

Fields

poolName :: Maybe Text

The name of the dedicated IP pool that the IP address is associated with.

ip :: Text

An IPv4 address.

warmupStatus :: WarmupStatus

The warm-up status of a dedicated IP address. The status can have one of the following values:

IN_PROGRESS – The IP address isn't ready to use because the dedicated IP warm-up process is ongoing.

DONE – The dedicated IP warm-up process is complete, and the IP address is ready to use.

warmupPercentage :: Int

Indicates how complete the dedicated IP warm-up process is. When this value equals 1, the address has completed the warm-up process and is ready for use.

newDedicatedIp

Arguments

:: Text
$sel:ip:DedicatedIp'

-> WarmupStatus
$sel:warmupStatus:DedicatedIp'

-> Int
$sel:warmupPercentage:DedicatedIp'

-> DedicatedIp

Create a value of DedicatedIp with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:poolName:DedicatedIp', dedicatedIp_poolName - The name of the dedicated IP pool that the IP address is associated with.

$sel:ip:DedicatedIp', dedicatedIp_ip - An IPv4 address.

$sel:warmupStatus:DedicatedIp', dedicatedIp_warmupStatus - The warm-up status of a dedicated IP address. The status can have one of the following values:

IN_PROGRESS – The IP address isn't ready to use because the dedicated IP warm-up process is ongoing.

DONE – The dedicated IP warm-up process is complete, and the IP address is ready to use.

$sel:warmupPercentage:DedicatedIp', dedicatedIp_warmupPercentage - Indicates how complete the dedicated IP warm-up process is. When this value equals 1, the address has completed the warm-up process and is ready for use.

dedicatedIp_poolName :: Lens' DedicatedIp (Maybe Text)

The name of the dedicated IP pool that the IP address is associated with.

dedicatedIp_ip :: Lens' DedicatedIp Text

An IPv4 address.

dedicatedIp_warmupStatus :: Lens' DedicatedIp WarmupStatus

The warm-up status of a dedicated IP address. The status can have one of the following values:

IN_PROGRESS – The IP address isn't ready to use because the dedicated IP warm-up process is ongoing.

DONE – The dedicated IP warm-up process is complete, and the IP address is ready to use.

dedicatedIp_warmupPercentage :: Lens' DedicatedIp Int

Indicates how complete the dedicated IP warm-up process is. When this value equals 1, the address has completed the warm-up process and is ready for use.