Network.AWS.SESv2.Types.Destination

Description

Documentation

data Destination

An object that describes the recipients for an email.

Amazon SES does not support the SMTPUTF8 extension, as described in RFC6531. For this reason, the local part of a destination email address (the part of the email address that precedes the @ sign) may only contain 7-bit ASCII characters. If the domain part of an address (the part after the @ sign) contains non-ASCII characters, they must be encoded using Punycode, as described in RFC3492.

See: newDestination smart constructor.

Constructors

Destination'

Fields

bccAddresses :: Maybe [Text]

An array that contains the email addresses of the "BCC" (blind carbon copy) recipients for the email.

ccAddresses :: Maybe [Text]

An array that contains the email addresses of the "CC" (carbon copy) recipients for the email.

toAddresses :: Maybe [Text]

An array that contains the email addresses of the "To" recipients for the email.

newDestination :: Destination

Create a value of Destination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bccAddresses:Destination', destination_bccAddresses - An array that contains the email addresses of the "BCC" (blind carbon copy) recipients for the email.

$sel:ccAddresses:Destination', destination_ccAddresses - An array that contains the email addresses of the "CC" (carbon copy) recipients for the email.

$sel:toAddresses:Destination', destination_toAddresses - An array that contains the email addresses of the "To" recipients for the email.

destination_bccAddresses :: Lens' Destination (Maybe [Text])

An array that contains the email addresses of the "BCC" (blind carbon copy) recipients for the email.

destination_ccAddresses :: Lens' Destination (Maybe [Text])

An array that contains the email addresses of the "CC" (carbon copy) recipients for the email.

destination_toAddresses :: Lens' Destination (Maybe [Text])

An array that contains the email addresses of the "To" recipients for the email.