Network.AWS.SESv2.Types.DomainDeliverabilityTrackingOption

Description

Documentation

data DomainDeliverabilityTrackingOption

An object that contains information about the Deliverability dashboard subscription for a verified domain that you use to send email and currently has an active Deliverability dashboard subscription. If a Deliverability dashboard subscription is active for a domain, you gain access to reputation, inbox placement, and other metrics for the domain.

See: newDomainDeliverabilityTrackingOption smart constructor.

Constructors

DomainDeliverabilityTrackingOption'

Fields

subscriptionStartDate :: Maybe POSIX

The date, in Unix time format, when you enabled the Deliverability dashboard for the domain.

domain :: Maybe Text

A verified domain that’s associated with your AWS account and currently has an active Deliverability dashboard subscription.

inboxPlacementTrackingOption :: Maybe InboxPlacementTrackingOption

An object that contains information about the inbox placement data settings for the domain.

newDomainDeliverabilityTrackingOption :: DomainDeliverabilityTrackingOption

Create a value of DomainDeliverabilityTrackingOption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscriptionStartDate:DomainDeliverabilityTrackingOption', domainDeliverabilityTrackingOption_subscriptionStartDate - The date, in Unix time format, when you enabled the Deliverability dashboard for the domain.

$sel:domain:DomainDeliverabilityTrackingOption', domainDeliverabilityTrackingOption_domain - A verified domain that’s associated with your AWS account and currently has an active Deliverability dashboard subscription.

$sel:inboxPlacementTrackingOption:DomainDeliverabilityTrackingOption', domainDeliverabilityTrackingOption_inboxPlacementTrackingOption - An object that contains information about the inbox placement data settings for the domain.

domainDeliverabilityTrackingOption_subscriptionStartDate :: Lens' DomainDeliverabilityTrackingOption (Maybe UTCTime)

The date, in Unix time format, when you enabled the Deliverability dashboard for the domain.

domainDeliverabilityTrackingOption_domain :: Lens' DomainDeliverabilityTrackingOption (Maybe Text)

A verified domain that’s associated with your AWS account and currently has an active Deliverability dashboard subscription.

domainDeliverabilityTrackingOption_inboxPlacementTrackingOption :: Lens' DomainDeliverabilityTrackingOption (Maybe InboxPlacementTrackingOption)

An object that contains information about the inbox placement data settings for the domain.