Network.AWS.SESv2.Types.EmailContent

Description

Documentation

data EmailContent

An object that defines the entire content of the email, including the message headers and the body content. You can create a simple email message, in which you specify the subject and the text and HTML versions of the message body. You can also create raw messages, in which you specify a complete MIME-formatted message. Raw messages can include attachments and custom headers.

See: newEmailContent smart constructor.

Constructors

EmailContent'

Fields

raw :: Maybe RawMessage

The raw email message. The message has to meet the following criteria:

The message has to contain a header and a body, separated by one blank line.

All of the required header fields must be present in the message.

Each part of a multipart MIME message must be formatted properly.

If you include attachments, they must be in a file format that the Amazon SES API v2 supports.

The entire message must be Base64 encoded.

If any of the MIME parts in your message contain content that is outside of the 7-bit ASCII character range, you should encode that content to ensure that recipients' email clients render the message properly.

The length of any single line of text in the message can't exceed 1,000 characters. This restriction is defined in RFC 5321.

simple :: Maybe Message

The simple email message. The message consists of a subject and a message body.

template :: Maybe Template

The template to use for the email message.

newEmailContent :: EmailContent

Create a value of EmailContent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:raw:EmailContent', emailContent_raw - The raw email message. The message has to meet the following criteria:

The message has to contain a header and a body, separated by one blank line.

All of the required header fields must be present in the message.

Each part of a multipart MIME message must be formatted properly.

If you include attachments, they must be in a file format that the Amazon SES API v2 supports.

The entire message must be Base64 encoded.

If any of the MIME parts in your message contain content that is outside of the 7-bit ASCII character range, you should encode that content to ensure that recipients' email clients render the message properly.

The length of any single line of text in the message can't exceed 1,000 characters. This restriction is defined in RFC 5321.

$sel:simple:EmailContent', emailContent_simple - The simple email message. The message consists of a subject and a message body.

$sel:template:EmailContent', emailContent_template - The template to use for the email message.

emailContent_raw :: Lens' EmailContent (Maybe RawMessage)

The raw email message. The message has to meet the following criteria:

The message has to contain a header and a body, separated by one blank line.

All of the required header fields must be present in the message.

Each part of a multipart MIME message must be formatted properly.

If you include attachments, they must be in a file format that the Amazon SES API v2 supports.

The entire message must be Base64 encoded.

If any of the MIME parts in your message contain content that is outside of the 7-bit ASCII character range, you should encode that content to ensure that recipients' email clients render the message properly.

The length of any single line of text in the message can't exceed 1,000 characters. This restriction is defined in RFC 5321.

emailContent_simple :: Lens' EmailContent (Maybe Message)

The simple email message. The message consists of a subject and a message body.

emailContent_template :: Lens' EmailContent (Maybe Template)

The template to use for the email message.