Network.AWS.SESv2.Types.EventDestination

Description

Documentation

data EventDestination

In the Amazon SES API v2, events include message sends, deliveries, opens, clicks, bounces, complaints and delivery delays. /Event destinations/ are places that you can send information about these events to. For example, you can send event data to Amazon SNS to receive notifications when you receive bounces or complaints, or you can use Amazon Kinesis Data Firehose to stream data to Amazon S3 for long-term storage.

See: newEventDestination smart constructor.

Constructors

EventDestination'

Fields

cloudWatchDestination :: Maybe CloudWatchDestination

An object that defines an Amazon CloudWatch destination for email events. You can use Amazon CloudWatch to monitor and gain insights on your email sending metrics.

enabled :: Maybe Bool

If true, the event destination is enabled. When the event destination is enabled, the specified event types are sent to the destinations in this EventDestinationDefinition.

If false, the event destination is disabled. When the event destination is disabled, events aren't sent to the specified destinations.

pinpointDestination :: Maybe PinpointDestination

An object that defines an Amazon Pinpoint project destination for email events. You can send email event data to a Amazon Pinpoint project to view metrics using the Transactional Messaging dashboards that are built in to Amazon Pinpoint. For more information, see Transactional Messaging Charts in the Amazon Pinpoint User Guide.

kinesisFirehoseDestination :: Maybe KinesisFirehoseDestination

An object that defines an Amazon Kinesis Data Firehose destination for email events. You can use Amazon Kinesis Data Firehose to stream data to other services, such as Amazon S3 and Amazon Redshift.

snsDestination :: Maybe SnsDestination

An object that defines an Amazon SNS destination for email events. You can use Amazon SNS to send notification when certain email events occur.

name :: Text

A name that identifies the event destination.

matchingEventTypes :: [EventType]

The types of events that Amazon SES sends to the specified event destinations.

newEventDestination

Arguments

:: Text
$sel:name:EventDestination'

-> EventDestination

Create a value of EventDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cloudWatchDestination:EventDestination', eventDestination_cloudWatchDestination - An object that defines an Amazon CloudWatch destination for email events. You can use Amazon CloudWatch to monitor and gain insights on your email sending metrics.

$sel:enabled:EventDestination', eventDestination_enabled - If true, the event destination is enabled. When the event destination is enabled, the specified event types are sent to the destinations in this EventDestinationDefinition.

If false, the event destination is disabled. When the event destination is disabled, events aren't sent to the specified destinations.

$sel:pinpointDestination:EventDestination', eventDestination_pinpointDestination - An object that defines an Amazon Pinpoint project destination for email events. You can send email event data to a Amazon Pinpoint project to view metrics using the Transactional Messaging dashboards that are built in to Amazon Pinpoint. For more information, see Transactional Messaging Charts in the Amazon Pinpoint User Guide.

$sel:kinesisFirehoseDestination:EventDestination', eventDestination_kinesisFirehoseDestination - An object that defines an Amazon Kinesis Data Firehose destination for email events. You can use Amazon Kinesis Data Firehose to stream data to other services, such as Amazon S3 and Amazon Redshift.

$sel:snsDestination:EventDestination', eventDestination_snsDestination - An object that defines an Amazon SNS destination for email events. You can use Amazon SNS to send notification when certain email events occur.

$sel:name:EventDestination', eventDestination_name - A name that identifies the event destination.

$sel:matchingEventTypes:EventDestination', eventDestination_matchingEventTypes - The types of events that Amazon SES sends to the specified event destinations.

eventDestination_cloudWatchDestination :: Lens' EventDestination (Maybe CloudWatchDestination)

An object that defines an Amazon CloudWatch destination for email events. You can use Amazon CloudWatch to monitor and gain insights on your email sending metrics.

eventDestination_enabled :: Lens' EventDestination (Maybe Bool)

If true, the event destination is enabled. When the event destination is enabled, the specified event types are sent to the destinations in this EventDestinationDefinition.

If false, the event destination is disabled. When the event destination is disabled, events aren't sent to the specified destinations.

eventDestination_pinpointDestination :: Lens' EventDestination (Maybe PinpointDestination)

An object that defines an Amazon Pinpoint project destination for email events. You can send email event data to a Amazon Pinpoint project to view metrics using the Transactional Messaging dashboards that are built in to Amazon Pinpoint. For more information, see Transactional Messaging Charts in the Amazon Pinpoint User Guide.

eventDestination_kinesisFirehoseDestination :: Lens' EventDestination (Maybe KinesisFirehoseDestination)

An object that defines an Amazon Kinesis Data Firehose destination for email events. You can use Amazon Kinesis Data Firehose to stream data to other services, such as Amazon S3 and Amazon Redshift.

eventDestination_snsDestination :: Lens' EventDestination (Maybe SnsDestination)

An object that defines an Amazon SNS destination for email events. You can use Amazon SNS to send notification when certain email events occur.

eventDestination_name :: Lens' EventDestination Text

A name that identifies the event destination.

eventDestination_matchingEventTypes :: Lens' EventDestination [EventType]

The types of events that Amazon SES sends to the specified event destinations.