Network.AWS.SESv2.Types.MessageTag

Description

Documentation

data MessageTag

Contains the name and value of a tag that you apply to an email. You can use message tags when you publish email sending events.

See: newMessageTag smart constructor.

Constructors

MessageTag'

Fields

name :: Text

The name of the message tag. The message tag name has to meet the following criteria:

It can only contain ASCII letters (a–z, A–Z), numbers (0–9), underscores (_), or dashes (-).

It can contain no more than 256 characters.

value :: Text

The value of the message tag. The message tag value has to meet the following criteria:

It can only contain ASCII letters (a–z, A–Z), numbers (0–9), underscores (_), or dashes (-).

It can contain no more than 256 characters.

newMessageTag

Arguments

:: Text
$sel:name:MessageTag'

-> Text
$sel:value:MessageTag'

-> MessageTag

Create a value of MessageTag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:MessageTag', messageTag_name - The name of the message tag. The message tag name has to meet the following criteria:

It can only contain ASCII letters (a–z, A–Z), numbers (0–9), underscores (_), or dashes (-).

It can contain no more than 256 characters.

$sel:value:MessageTag', messageTag_value - The value of the message tag. The message tag value has to meet the following criteria:

It can only contain ASCII letters (a–z, A–Z), numbers (0–9), underscores (_), or dashes (-).

It can contain no more than 256 characters.

messageTag_name :: Lens' MessageTag Text

The name of the message tag. The message tag name has to meet the following criteria:

It can only contain ASCII letters (a–z, A–Z), numbers (0–9), underscores (_), or dashes (-).

It can contain no more than 256 characters.

messageTag_value :: Lens' MessageTag Text

The value of the message tag. The message tag value has to meet the following criteria:

It can only contain ASCII letters (a–z, A–Z), numbers (0–9), underscores (_), or dashes (-).

It can contain no more than 256 characters.