Network.AWS.SESv2.Types.OverallVolume

Description

Documentation

data OverallVolume

An object that contains information about email that was sent from the selected domain.

See: newOverallVolume smart constructor.

Constructors

OverallVolume'

Fields

domainIspPlacements :: Maybe [DomainIspPlacement]

An object that contains inbox and junk mail placement metrics for individual email providers.

volumeStatistics :: Maybe VolumeStatistics

An object that contains information about the numbers of messages that arrived in recipients' inboxes and junk mail folders.

readRatePercent :: Maybe Double

The percentage of emails that were sent from the domain that were read by their recipients.

newOverallVolume :: OverallVolume

Create a value of OverallVolume with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainIspPlacements:OverallVolume', overallVolume_domainIspPlacements - An object that contains inbox and junk mail placement metrics for individual email providers.

$sel:volumeStatistics:OverallVolume', overallVolume_volumeStatistics - An object that contains information about the numbers of messages that arrived in recipients' inboxes and junk mail folders.

$sel:readRatePercent:OverallVolume', overallVolume_readRatePercent - The percentage of emails that were sent from the domain that were read by their recipients.

overallVolume_domainIspPlacements :: Lens' OverallVolume (Maybe [DomainIspPlacement])

An object that contains inbox and junk mail placement metrics for individual email providers.

overallVolume_volumeStatistics :: Lens' OverallVolume (Maybe VolumeStatistics)

An object that contains information about the numbers of messages that arrived in recipients' inboxes and junk mail folders.

overallVolume_readRatePercent :: Lens' OverallVolume (Maybe Double)

The percentage of emails that were sent from the domain that were read by their recipients.