Network.AWS.SESv2.Types.PlacementStatistics

Description

Documentation

data PlacementStatistics

An object that contains inbox placement data for an email provider.

See: newPlacementStatistics smart constructor.

Constructors

PlacementStatistics'

Fields

spamPercentage :: Maybe Double

The percentage of emails that arrived in recipients' spam or junk mail folders during the predictive inbox placement test.

dkimPercentage :: Maybe Double

The percentage of emails that were authenticated by using DomainKeys Identified Mail (DKIM) during the predictive inbox placement test.

spfPercentage :: Maybe Double

The percentage of emails that were authenticated by using Sender Policy Framework (SPF) during the predictive inbox placement test.

inboxPercentage :: Maybe Double

The percentage of emails that arrived in recipients' inboxes during the predictive inbox placement test.

missingPercentage :: Maybe Double

The percentage of emails that didn't arrive in recipients' inboxes at all during the predictive inbox placement test.

newPlacementStatistics :: PlacementStatistics

Create a value of PlacementStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:spamPercentage:PlacementStatistics', placementStatistics_spamPercentage - The percentage of emails that arrived in recipients' spam or junk mail folders during the predictive inbox placement test.

$sel:dkimPercentage:PlacementStatistics', placementStatistics_dkimPercentage - The percentage of emails that were authenticated by using DomainKeys Identified Mail (DKIM) during the predictive inbox placement test.

$sel:spfPercentage:PlacementStatistics', placementStatistics_spfPercentage - The percentage of emails that were authenticated by using Sender Policy Framework (SPF) during the predictive inbox placement test.

$sel:inboxPercentage:PlacementStatistics', placementStatistics_inboxPercentage - The percentage of emails that arrived in recipients' inboxes during the predictive inbox placement test.

$sel:missingPercentage:PlacementStatistics', placementStatistics_missingPercentage - The percentage of emails that didn't arrive in recipients' inboxes at all during the predictive inbox placement test.

placementStatistics_spamPercentage :: Lens' PlacementStatistics (Maybe Double)

The percentage of emails that arrived in recipients' spam or junk mail folders during the predictive inbox placement test.

placementStatistics_dkimPercentage :: Lens' PlacementStatistics (Maybe Double)

The percentage of emails that were authenticated by using DomainKeys Identified Mail (DKIM) during the predictive inbox placement test.

placementStatistics_spfPercentage :: Lens' PlacementStatistics (Maybe Double)

The percentage of emails that were authenticated by using Sender Policy Framework (SPF) during the predictive inbox placement test.

placementStatistics_inboxPercentage :: Lens' PlacementStatistics (Maybe Double)

The percentage of emails that arrived in recipients' inboxes during the predictive inbox placement test.

placementStatistics_missingPercentage :: Lens' PlacementStatistics (Maybe Double)

The percentage of emails that didn't arrive in recipients' inboxes at all during the predictive inbox placement test.