Network.AWS.SESv2.Types.ReputationOptions

Description

Documentation

data ReputationOptions

Enable or disable collection of reputation metrics for emails that you send using this configuration set in the current AWS Region.

See: newReputationOptions smart constructor.

Constructors

ReputationOptions'

Fields

reputationMetricsEnabled :: Maybe Bool

If true, tracking of reputation metrics is enabled for the configuration set. If false, tracking of reputation metrics is disabled for the configuration set.

lastFreshStart :: Maybe POSIX

The date and time (in Unix time) when the reputation metrics were last given a fresh start. When your account is given a fresh start, your reputation metrics are calculated starting from the date of the fresh start.

newReputationOptions :: ReputationOptions

Create a value of ReputationOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reputationMetricsEnabled:ReputationOptions', reputationOptions_reputationMetricsEnabled - If true, tracking of reputation metrics is enabled for the configuration set. If false, tracking of reputation metrics is disabled for the configuration set.

$sel:lastFreshStart:ReputationOptions', reputationOptions_lastFreshStart - The date and time (in Unix time) when the reputation metrics were last given a fresh start. When your account is given a fresh start, your reputation metrics are calculated starting from the date of the fresh start.

reputationOptions_reputationMetricsEnabled :: Lens' ReputationOptions (Maybe Bool)

If true, tracking of reputation metrics is enabled for the configuration set. If false, tracking of reputation metrics is disabled for the configuration set.

reputationOptions_lastFreshStart :: Lens' ReputationOptions (Maybe UTCTime)

The date and time (in Unix time) when the reputation metrics were last given a fresh start. When your account is given a fresh start, your reputation metrics are calculated starting from the date of the fresh start.