Network.AWS.SESv2.Types.TrackingOptions

Description

Documentation

data TrackingOptions

An object that defines the tracking options for a configuration set. When you use the Amazon SES API v2 to send an email, it contains an invisible image that's used to track when recipients open your email. If your email contains links, those links are changed slightly in order to track when recipients click them.

These images and links include references to a domain operated by AWS. You can optionally configure the Amazon SES to use a domain that you operate for these images and links.

See: newTrackingOptions smart constructor.

Constructors

TrackingOptions'

Fields

customRedirectDomain :: Text

The domain that you want to use for tracking open and click events.

newTrackingOptions

Arguments

:: Text
$sel:customRedirectDomain:TrackingOptions'

-> TrackingOptions

Create a value of TrackingOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customRedirectDomain:TrackingOptions', trackingOptions_customRedirectDomain - The domain that you want to use for tracking open and click events.

trackingOptions_customRedirectDomain :: Lens' TrackingOptions Text

The domain that you want to use for tracking open and click events.