Network.AWS.SESv2.UpdateEmailTemplate

Description

Updates an email template. Email templates enable you to send personalized email to one or more destinations in a single API operation. For more information, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data UpdateEmailTemplate

Represents a request to update an email template. For more information, see the Amazon SES Developer Guide.

See: newUpdateEmailTemplate smart constructor.

Constructors

UpdateEmailTemplate'

Fields

templateName :: Text

The name of the template you want to update.

templateContent :: EmailTemplateContent

The content of the email template, composed of a subject line, an HTML part, and a text-only part.

newUpdateEmailTemplate

Arguments

:: Text
$sel:templateName:UpdateEmailTemplate'

-> EmailTemplateContent
$sel:templateContent:UpdateEmailTemplate'

-> UpdateEmailTemplate

Create a value of UpdateEmailTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:UpdateEmailTemplate', updateEmailTemplate_templateName - The name of the template you want to update.

$sel:templateContent:UpdateEmailTemplate', updateEmailTemplate_templateContent - The content of the email template, composed of a subject line, an HTML part, and a text-only part.

Request Lenses

updateEmailTemplate_templateName :: Lens' UpdateEmailTemplate Text

The name of the template you want to update.

updateEmailTemplate_templateContent :: Lens' UpdateEmailTemplate EmailTemplateContent

The content of the email template, composed of a subject line, an HTML part, and a text-only part.

Destructuring the Response

data UpdateEmailTemplateResponse

If the action is successful, the service sends back an HTTP 200 response with an empty HTTP body.

See: newUpdateEmailTemplateResponse smart constructor.

Constructors

UpdateEmailTemplateResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateEmailTemplateResponse

Arguments

:: Int
$sel:httpStatus:UpdateEmailTemplateResponse'

-> UpdateEmailTemplateResponse

Create a value of UpdateEmailTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateEmailTemplateResponse', updateEmailTemplateResponse_httpStatus - The response's http status code.

Response Lenses

updateEmailTemplateResponse_httpStatus :: Lens' UpdateEmailTemplateResponse Int

The response's http status code.