Network.AWS.SMS.CreateReplicationJob

Description

Creates a replication job. The replication job schedules periodic replication runs to replicate your server to AWS. Each replication run creates an Amazon Machine Image (AMI).

Creating a Request

data CreateReplicationJob

See: newCreateReplicationJob smart constructor.

Constructors

CreateReplicationJob'

Fields

frequency :: Maybe Int

The time between consecutive replication runs, in hours.

numberOfRecentAmisToKeep :: Maybe Int

The maximum number of SMS-created AMIs to retain. The oldest is deleted after the maximum number is reached and a new AMI is created.

licenseType :: Maybe LicenseType

The license type to be used for the AMI created by a successful replication run.

roleName :: Maybe Text

The name of the IAM role to be used by the AWS SMS.

encrypted :: Maybe Bool

Indicates whether the replication job produces encrypted AMIs.

kmsKeyId :: Maybe Text

The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is true but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

runOnce :: Maybe Bool

Indicates whether to run the replication job one time.

description :: Maybe Text

The description of the replication job.

serverId :: Text

The ID of the server.

seedReplicationTime :: POSIX

The seed replication time.

newCreateReplicationJob

Arguments

:: Text
$sel:serverId:CreateReplicationJob'

-> UTCTime
$sel:seedReplicationTime:CreateReplicationJob'

-> CreateReplicationJob

Create a value of CreateReplicationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:frequency:CreateReplicationJob', createReplicationJob_frequency - The time between consecutive replication runs, in hours.

$sel:numberOfRecentAmisToKeep:CreateReplicationJob', createReplicationJob_numberOfRecentAmisToKeep - The maximum number of SMS-created AMIs to retain. The oldest is deleted after the maximum number is reached and a new AMI is created.

$sel:licenseType:CreateReplicationJob', createReplicationJob_licenseType - The license type to be used for the AMI created by a successful replication run.

$sel:roleName:CreateReplicationJob', createReplicationJob_roleName - The name of the IAM role to be used by the AWS SMS.

$sel:encrypted:CreateReplicationJob', createReplicationJob_encrypted - Indicates whether the replication job produces encrypted AMIs.

$sel:kmsKeyId:CreateReplicationJob', createReplicationJob_kmsKeyId - The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is true but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

$sel:runOnce:CreateReplicationJob', createReplicationJob_runOnce - Indicates whether to run the replication job one time.

$sel:description:CreateReplicationJob', createReplicationJob_description - The description of the replication job.

$sel:serverId:CreateReplicationJob', createReplicationJob_serverId - The ID of the server.

$sel:seedReplicationTime:CreateReplicationJob', createReplicationJob_seedReplicationTime - The seed replication time.

Request Lenses

createReplicationJob_frequency :: Lens' CreateReplicationJob (Maybe Int)

The time between consecutive replication runs, in hours.

createReplicationJob_numberOfRecentAmisToKeep :: Lens' CreateReplicationJob (Maybe Int)

The maximum number of SMS-created AMIs to retain. The oldest is deleted after the maximum number is reached and a new AMI is created.

createReplicationJob_licenseType :: Lens' CreateReplicationJob (Maybe LicenseType)

The license type to be used for the AMI created by a successful replication run.

createReplicationJob_roleName :: Lens' CreateReplicationJob (Maybe Text)

The name of the IAM role to be used by the AWS SMS.

createReplicationJob_encrypted :: Lens' CreateReplicationJob (Maybe Bool)

Indicates whether the replication job produces encrypted AMIs.

createReplicationJob_kmsKeyId :: Lens' CreateReplicationJob (Maybe Text)

The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is true but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

createReplicationJob_runOnce :: Lens' CreateReplicationJob (Maybe Bool)

Indicates whether to run the replication job one time.

createReplicationJob_description :: Lens' CreateReplicationJob (Maybe Text)

The description of the replication job.

createReplicationJob_serverId :: Lens' CreateReplicationJob Text

The ID of the server.

createReplicationJob_seedReplicationTime :: Lens' CreateReplicationJob UTCTime

The seed replication time.

Destructuring the Response

data CreateReplicationJobResponse

See: newCreateReplicationJobResponse smart constructor.

Constructors

CreateReplicationJobResponse'

Fields

replicationJobId :: Maybe Text

The unique identifier of the replication job.

httpStatus :: Int

The response's http status code.

newCreateReplicationJobResponse

Arguments

:: Int
$sel:httpStatus:CreateReplicationJobResponse'

-> CreateReplicationJobResponse

Create a value of CreateReplicationJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationJobId:CreateReplicationJobResponse', createReplicationJobResponse_replicationJobId - The unique identifier of the replication job.

$sel:httpStatus:CreateReplicationJobResponse', createReplicationJobResponse_httpStatus - The response's http status code.

Response Lenses

createReplicationJobResponse_replicationJobId :: Lens' CreateReplicationJobResponse (Maybe Text)

The unique identifier of the replication job.

createReplicationJobResponse_httpStatus :: Lens' CreateReplicationJobResponse Int

The response's http status code.