Network.AWS.SMS.DeleteAppReplicationConfiguration

Description

Deletes the replication configuration for the specified application.

Creating a Request

data DeleteAppReplicationConfiguration

See: newDeleteAppReplicationConfiguration smart constructor.

Constructors

DeleteAppReplicationConfiguration'

Fields

appId :: Maybe Text

The ID of the application.

newDeleteAppReplicationConfiguration :: DeleteAppReplicationConfiguration

Create a value of DeleteAppReplicationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appId:DeleteAppReplicationConfiguration', deleteAppReplicationConfiguration_appId - The ID of the application.

Request Lenses

deleteAppReplicationConfiguration_appId :: Lens' DeleteAppReplicationConfiguration (Maybe Text)

The ID of the application.

Destructuring the Response

data DeleteAppReplicationConfigurationResponse

See: newDeleteAppReplicationConfigurationResponse smart constructor.

Constructors

DeleteAppReplicationConfigurationResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteAppReplicationConfigurationResponse

Arguments

:: Int
$sel:httpStatus:DeleteAppReplicationConfigurationResponse'

-> DeleteAppReplicationConfigurationResponse

Create a value of DeleteAppReplicationConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteAppReplicationConfigurationResponse', deleteAppReplicationConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

deleteAppReplicationConfigurationResponse_httpStatus :: Lens' DeleteAppReplicationConfigurationResponse Int

The response's http status code.