Network.AWS.SMS.DeleteReplicationJob

Description

Deletes the specified replication job.

After you delete a replication job, there are no further replication runs. AWS deletes the contents of the Amazon S3 bucket used to store AWS SMS artifacts. The AMIs created by the replication runs are not deleted.

Creating a Request

data DeleteReplicationJob

See: newDeleteReplicationJob smart constructor.

Constructors

DeleteReplicationJob'

Fields

replicationJobId :: Text

The ID of the replication job.

newDeleteReplicationJob

Arguments

:: Text
$sel:replicationJobId:DeleteReplicationJob'

-> DeleteReplicationJob

Create a value of DeleteReplicationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicationJobId:DeleteReplicationJob', deleteReplicationJob_replicationJobId - The ID of the replication job.

Request Lenses

deleteReplicationJob_replicationJobId :: Lens' DeleteReplicationJob Text

The ID of the replication job.

Destructuring the Response

data DeleteReplicationJobResponse

See: newDeleteReplicationJobResponse smart constructor.

Constructors

DeleteReplicationJobResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteReplicationJobResponse

Arguments

:: Int
$sel:httpStatus:DeleteReplicationJobResponse'

-> DeleteReplicationJobResponse

Create a value of DeleteReplicationJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteReplicationJobResponse', deleteReplicationJobResponse_httpStatus - The response's http status code.

Response Lenses

deleteReplicationJobResponse_httpStatus :: Lens' DeleteReplicationJobResponse Int

The response's http status code.