Network.AWS.SMS.GetAppValidationConfiguration

Description

Retrieves information about a configuration for validating an application.

Creating a Request

data GetAppValidationConfiguration

See: newGetAppValidationConfiguration smart constructor.

Constructors

GetAppValidationConfiguration'

Fields

appId :: Text

The ID of the application.

newGetAppValidationConfiguration

Arguments

:: Text
$sel:appId:GetAppValidationConfiguration'

-> GetAppValidationConfiguration

Create a value of GetAppValidationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appId:GetAppValidationConfiguration', getAppValidationConfiguration_appId - The ID of the application.

Request Lenses

getAppValidationConfiguration_appId :: Lens' GetAppValidationConfiguration Text

The ID of the application.

Destructuring the Response

data GetAppValidationConfigurationResponse

See: newGetAppValidationConfigurationResponse smart constructor.

Constructors

GetAppValidationConfigurationResponse'

Fields

appValidationConfigurations :: Maybe [AppValidationConfiguration]

The configuration for application validation.

serverGroupValidationConfigurations :: Maybe [ServerGroupValidationConfiguration]

The configuration for instance validation.

httpStatus :: Int

The response's http status code.

newGetAppValidationConfigurationResponse

Arguments

:: Int
$sel:httpStatus:GetAppValidationConfigurationResponse'

-> GetAppValidationConfigurationResponse

Create a value of GetAppValidationConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appValidationConfigurations:GetAppValidationConfigurationResponse', getAppValidationConfigurationResponse_appValidationConfigurations - The configuration for application validation.

$sel:serverGroupValidationConfigurations:GetAppValidationConfigurationResponse', getAppValidationConfigurationResponse_serverGroupValidationConfigurations - The configuration for instance validation.

$sel:httpStatus:GetAppValidationConfigurationResponse', getAppValidationConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

getAppValidationConfigurationResponse_appValidationConfigurations :: Lens' GetAppValidationConfigurationResponse (Maybe [AppValidationConfiguration])

The configuration for application validation.

getAppValidationConfigurationResponse_serverGroupValidationConfigurations :: Lens' GetAppValidationConfigurationResponse (Maybe [ServerGroupValidationConfiguration])

The configuration for instance validation.

getAppValidationConfigurationResponse_httpStatus :: Lens' GetAppValidationConfigurationResponse Int

The response's http status code.