Network.AWS.SMS.GetAppValidationOutput

Description

Retrieves output from validating an application.

Creating a Request

data GetAppValidationOutput

See: newGetAppValidationOutput smart constructor.

Constructors

GetAppValidationOutput'

Fields

appId :: Text

The ID of the application.

newGetAppValidationOutput

Arguments

:: Text
$sel:appId:GetAppValidationOutput'

-> GetAppValidationOutput

Create a value of GetAppValidationOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appId:GetAppValidationOutput', getAppValidationOutput_appId - The ID of the application.

Request Lenses

getAppValidationOutput_appId :: Lens' GetAppValidationOutput Text

The ID of the application.

Destructuring the Response

data GetAppValidationOutputResponse

See: newGetAppValidationOutputResponse smart constructor.

Constructors

GetAppValidationOutputResponse'

Fields

validationOutputList :: Maybe [ValidationOutput]

The validation output.

httpStatus :: Int

The response's http status code.

newGetAppValidationOutputResponse

Arguments

:: Int
$sel:httpStatus:GetAppValidationOutputResponse'

-> GetAppValidationOutputResponse

Create a value of GetAppValidationOutputResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:validationOutputList:GetAppValidationOutputResponse', getAppValidationOutputResponse_validationOutputList - The validation output.

$sel:httpStatus:GetAppValidationOutputResponse', getAppValidationOutputResponse_httpStatus - The response's http status code.

Response Lenses

getAppValidationOutputResponse_validationOutputList :: Lens' GetAppValidationOutputResponse (Maybe [ValidationOutput])

The validation output.

getAppValidationOutputResponse_httpStatus :: Lens' GetAppValidationOutputResponse Int

The response's http status code.