Network.AWS.SMS.GetConnectors

Description

Describes the connectors registered with the AWS SMS.

This operation returns paginated results.

Creating a Request

data GetConnectors

See: newGetConnectors smart constructor.

Constructors

GetConnectors'

Fields

nextToken :: Maybe Text

The token for the next set of results.

maxResults :: Maybe Int

The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned NextToken value.

newGetConnectors :: GetConnectors

Create a value of GetConnectors with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetConnectors', getConnectors_nextToken - The token for the next set of results.

$sel:maxResults:GetConnectors', getConnectors_maxResults - The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned NextToken value.

Request Lenses

getConnectors_nextToken :: Lens' GetConnectors (Maybe Text)

The token for the next set of results.

getConnectors_maxResults :: Lens' GetConnectors (Maybe Int)

The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned NextToken value.

Destructuring the Response

data GetConnectorsResponse

See: newGetConnectorsResponse smart constructor.

Constructors

GetConnectorsResponse'

Fields

nextToken :: Maybe Text

The token required to retrieve the next set of results. This value is null when there are no more results to return.

connectorList :: Maybe [Connector]

Information about the registered connectors.

httpStatus :: Int

The response's http status code.

newGetConnectorsResponse

Arguments

:: Int
$sel:httpStatus:GetConnectorsResponse'

-> GetConnectorsResponse

Create a value of GetConnectorsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetConnectors', getConnectorsResponse_nextToken - The token required to retrieve the next set of results. This value is null when there are no more results to return.

$sel:connectorList:GetConnectorsResponse', getConnectorsResponse_connectorList - Information about the registered connectors.

$sel:httpStatus:GetConnectorsResponse', getConnectorsResponse_httpStatus - The response's http status code.

Response Lenses

getConnectorsResponse_nextToken :: Lens' GetConnectorsResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null when there are no more results to return.

getConnectorsResponse_connectorList :: Lens' GetConnectorsResponse (Maybe [Connector])

Information about the registered connectors.

getConnectorsResponse_httpStatus :: Lens' GetConnectorsResponse Int

The response's http status code.