Network.AWS.SMS.GetReplicationJobs

Description

Describes the specified replication job or all of your replication jobs.

This operation returns paginated results.

Creating a Request

data GetReplicationJobs

See: newGetReplicationJobs smart constructor.

Constructors

GetReplicationJobs'

Fields

nextToken :: Maybe Text

The token for the next set of results.

maxResults :: Maybe Int

The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned NextToken value.

replicationJobId :: Maybe Text

The ID of the replication job.

newGetReplicationJobs :: GetReplicationJobs

Create a value of GetReplicationJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetReplicationJobs', getReplicationJobs_nextToken - The token for the next set of results.

$sel:maxResults:GetReplicationJobs', getReplicationJobs_maxResults - The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned NextToken value.

$sel:replicationJobId:GetReplicationJobs', getReplicationJobs_replicationJobId - The ID of the replication job.

Request Lenses

getReplicationJobs_nextToken :: Lens' GetReplicationJobs (Maybe Text)

The token for the next set of results.

getReplicationJobs_maxResults :: Lens' GetReplicationJobs (Maybe Int)

The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned NextToken value.

getReplicationJobs_replicationJobId :: Lens' GetReplicationJobs (Maybe Text)

The ID of the replication job.

Destructuring the Response

data GetReplicationJobsResponse

See: newGetReplicationJobsResponse smart constructor.

Constructors

GetReplicationJobsResponse'

Fields

nextToken :: Maybe Text

The token required to retrieve the next set of results. This value is null when there are no more results to return.

replicationJobList :: Maybe [ReplicationJob]

Information about the replication jobs.

httpStatus :: Int

The response's http status code.

newGetReplicationJobsResponse

Arguments

:: Int
$sel:httpStatus:GetReplicationJobsResponse'

-> GetReplicationJobsResponse

Create a value of GetReplicationJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetReplicationJobs', getReplicationJobsResponse_nextToken - The token required to retrieve the next set of results. This value is null when there are no more results to return.

$sel:replicationJobList:GetReplicationJobsResponse', getReplicationJobsResponse_replicationJobList - Information about the replication jobs.

$sel:httpStatus:GetReplicationJobsResponse', getReplicationJobsResponse_httpStatus - The response's http status code.

Response Lenses

getReplicationJobsResponse_nextToken :: Lens' GetReplicationJobsResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null when there are no more results to return.

getReplicationJobsResponse_replicationJobList :: Lens' GetReplicationJobsResponse (Maybe [ReplicationJob])

Information about the replication jobs.

getReplicationJobsResponse_httpStatus :: Lens' GetReplicationJobsResponse Int

The response's http status code.