Network.AWS.SMS.GetReplicationRuns

Description

Describes the replication runs for the specified replication job.

This operation returns paginated results.

Creating a Request

data GetReplicationRuns

See: newGetReplicationRuns smart constructor.

Constructors

GetReplicationRuns'

Fields

nextToken :: Maybe Text

The token for the next set of results.

maxResults :: Maybe Int

The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned NextToken value.

replicationJobId :: Text

The ID of the replication job.

newGetReplicationRuns

Arguments

:: Text
$sel:replicationJobId:GetReplicationRuns'

-> GetReplicationRuns

Create a value of GetReplicationRuns with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetReplicationRuns', getReplicationRuns_nextToken - The token for the next set of results.

$sel:maxResults:GetReplicationRuns', getReplicationRuns_maxResults - The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned NextToken value.

$sel:replicationJobId:GetReplicationRuns', getReplicationRuns_replicationJobId - The ID of the replication job.

Request Lenses

getReplicationRuns_nextToken :: Lens' GetReplicationRuns (Maybe Text)

The token for the next set of results.

getReplicationRuns_maxResults :: Lens' GetReplicationRuns (Maybe Int)

The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned NextToken value.

getReplicationRuns_replicationJobId :: Lens' GetReplicationRuns Text

The ID of the replication job.

Destructuring the Response

data GetReplicationRunsResponse

See: newGetReplicationRunsResponse smart constructor.

Constructors

GetReplicationRunsResponse'

Fields

nextToken :: Maybe Text

The token required to retrieve the next set of results. This value is null when there are no more results to return.

replicationJob :: Maybe ReplicationJob

Information about the replication job.

replicationRunList :: Maybe [ReplicationRun]

Information about the replication runs.

httpStatus :: Int

The response's http status code.

newGetReplicationRunsResponse

Arguments

:: Int
$sel:httpStatus:GetReplicationRunsResponse'

-> GetReplicationRunsResponse

Create a value of GetReplicationRunsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetReplicationRuns', getReplicationRunsResponse_nextToken - The token required to retrieve the next set of results. This value is null when there are no more results to return.

$sel:replicationJob:GetReplicationRunsResponse', getReplicationRunsResponse_replicationJob - Information about the replication job.

$sel:replicationRunList:GetReplicationRunsResponse', getReplicationRunsResponse_replicationRunList - Information about the replication runs.

$sel:httpStatus:GetReplicationRunsResponse', getReplicationRunsResponse_httpStatus - The response's http status code.

Response Lenses

getReplicationRunsResponse_nextToken :: Lens' GetReplicationRunsResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null when there are no more results to return.

getReplicationRunsResponse_replicationJob :: Lens' GetReplicationRunsResponse (Maybe ReplicationJob)

Information about the replication job.

getReplicationRunsResponse_replicationRunList :: Lens' GetReplicationRunsResponse (Maybe [ReplicationRun])

Information about the replication runs.

getReplicationRunsResponse_httpStatus :: Lens' GetReplicationRunsResponse Int

The response's http status code.