Network.AWS.SMS.ImportAppCatalog

Description

Allows application import from AWS Migration Hub.

Creating a Request

data ImportAppCatalog

See: newImportAppCatalog smart constructor.

Constructors

ImportAppCatalog'

Fields

roleName :: Maybe Text

The name of the service role. If you omit this parameter, we create a service-linked role for AWS Migration Hub in your account. Otherwise, the role that you provide must have the policy and trust policy described in the AWS Migration Hub User Guide.

newImportAppCatalog :: ImportAppCatalog

Create a value of ImportAppCatalog with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleName:ImportAppCatalog', importAppCatalog_roleName - The name of the service role. If you omit this parameter, we create a service-linked role for AWS Migration Hub in your account. Otherwise, the role that you provide must have the policy and trust policy described in the AWS Migration Hub User Guide.

Request Lenses

importAppCatalog_roleName :: Lens' ImportAppCatalog (Maybe Text)

The name of the service role. If you omit this parameter, we create a service-linked role for AWS Migration Hub in your account. Otherwise, the role that you provide must have the policy and trust policy described in the AWS Migration Hub User Guide.

Destructuring the Response

data ImportAppCatalogResponse

See: newImportAppCatalogResponse smart constructor.

Constructors

ImportAppCatalogResponse'

Fields

httpStatus :: Int

The response's http status code.

newImportAppCatalogResponse

Arguments

:: Int
$sel:httpStatus:ImportAppCatalogResponse'

-> ImportAppCatalogResponse

Create a value of ImportAppCatalogResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ImportAppCatalogResponse', importAppCatalogResponse_httpStatus - The response's http status code.

Response Lenses

importAppCatalogResponse_httpStatus :: Lens' ImportAppCatalogResponse Int

The response's http status code.