Network.AWS.SMS.Lens

Description

Operations

DeleteAppReplicationConfiguration

deleteAppReplicationConfiguration_appId :: Lens' DeleteAppReplicationConfiguration (Maybe Text)

The ID of the application.

deleteAppReplicationConfigurationResponse_httpStatus :: Lens' DeleteAppReplicationConfigurationResponse Int

The response's http status code.

PutAppReplicationConfiguration

putAppReplicationConfiguration_appId :: Lens' PutAppReplicationConfiguration (Maybe Text)

The ID of the application.

putAppReplicationConfiguration_serverGroupReplicationConfigurations :: Lens' PutAppReplicationConfiguration (Maybe [ServerGroupReplicationConfiguration])

Information about the replication configurations for server groups in the application.

putAppReplicationConfigurationResponse_httpStatus :: Lens' PutAppReplicationConfigurationResponse Int

The response's http status code.

DeleteServerCatalog

deleteServerCatalogResponse_httpStatus :: Lens' DeleteServerCatalogResponse Int

The response's http status code.

ImportAppCatalog

importAppCatalog_roleName :: Lens' ImportAppCatalog (Maybe Text)

The name of the service role. If you omit this parameter, we create a service-linked role for AWS Migration Hub in your account. Otherwise, the role that you provide must have the policy and trust policy described in the AWS Migration Hub User Guide.

importAppCatalogResponse_httpStatus :: Lens' ImportAppCatalogResponse Int

The response's http status code.

GetAppLaunchConfiguration

getAppLaunchConfiguration_appId :: Lens' GetAppLaunchConfiguration (Maybe Text)

The ID of the application.

getAppLaunchConfigurationResponse_serverGroupLaunchConfigurations :: Lens' GetAppLaunchConfigurationResponse (Maybe [ServerGroupLaunchConfiguration])

The launch configurations for server groups in this application.

getAppLaunchConfigurationResponse_autoLaunch :: Lens' GetAppLaunchConfigurationResponse (Maybe Bool)

Indicates whether the application is configured to launch automatically after replication is complete.

getAppLaunchConfigurationResponse_roleName :: Lens' GetAppLaunchConfigurationResponse (Maybe Text)

The name of the service role in the customer's account that AWS CloudFormation uses to launch the application.

getAppLaunchConfigurationResponse_appId :: Lens' GetAppLaunchConfigurationResponse (Maybe Text)

The ID of the application.

getAppLaunchConfigurationResponse_httpStatus :: Lens' GetAppLaunchConfigurationResponse Int

The response's http status code.

DeleteAppLaunchConfiguration

deleteAppLaunchConfiguration_appId :: Lens' DeleteAppLaunchConfiguration (Maybe Text)

The ID of the application.

deleteAppLaunchConfigurationResponse_httpStatus :: Lens' DeleteAppLaunchConfigurationResponse Int

The response's http status code.

StartAppReplication

startAppReplication_appId :: Lens' StartAppReplication (Maybe Text)

The ID of the application.

startAppReplicationResponse_httpStatus :: Lens' StartAppReplicationResponse Int

The response's http status code.

PutAppLaunchConfiguration

putAppLaunchConfiguration_serverGroupLaunchConfigurations :: Lens' PutAppLaunchConfiguration (Maybe [ServerGroupLaunchConfiguration])

Information about the launch configurations for server groups in the application.

putAppLaunchConfiguration_autoLaunch :: Lens' PutAppLaunchConfiguration (Maybe Bool)

Indicates whether the application is configured to launch automatically after replication is complete.

putAppLaunchConfiguration_roleName :: Lens' PutAppLaunchConfiguration (Maybe Text)

The name of service role in the customer's account that AWS CloudFormation uses to launch the application.

putAppLaunchConfiguration_appId :: Lens' PutAppLaunchConfiguration (Maybe Text)

The ID of the application.

putAppLaunchConfigurationResponse_httpStatus :: Lens' PutAppLaunchConfigurationResponse Int

The response's http status code.

GetReplicationRuns

getReplicationRuns_nextToken :: Lens' GetReplicationRuns (Maybe Text)

The token for the next set of results.

getReplicationRuns_maxResults :: Lens' GetReplicationRuns (Maybe Int)

The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned NextToken value.

getReplicationRuns_replicationJobId :: Lens' GetReplicationRuns Text

The ID of the replication job.

getReplicationRunsResponse_replicationJob :: Lens' GetReplicationRunsResponse (Maybe ReplicationJob)

Information about the replication job.

getReplicationRunsResponse_nextToken :: Lens' GetReplicationRunsResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null when there are no more results to return.

getReplicationRunsResponse_replicationRunList :: Lens' GetReplicationRunsResponse (Maybe [ReplicationRun])

Information about the replication runs.

getReplicationRunsResponse_httpStatus :: Lens' GetReplicationRunsResponse Int

The response's http status code.

TerminateApp

terminateApp_appId :: Lens' TerminateApp (Maybe Text)

The ID of the application.

terminateAppResponse_httpStatus :: Lens' TerminateAppResponse Int

The response's http status code.

ListApps

listApps_appIds :: Lens' ListApps (Maybe [Text])

The unique application IDs.

listApps_nextToken :: Lens' ListApps (Maybe Text)

The token for the next set of results.

listApps_maxResults :: Lens' ListApps (Maybe Int)

The maximum number of results to return in a single call. The default value is 100. To retrieve the remaining results, make another call with the returned NextToken value.

listAppsResponse_apps :: Lens' ListAppsResponse (Maybe [AppSummary])

The application summaries.

listAppsResponse_nextToken :: Lens' ListAppsResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null when there are no more results to return.

listAppsResponse_httpStatus :: Lens' ListAppsResponse Int

The response's http status code.

GetServers

getServers_vmServerAddressList :: Lens' GetServers (Maybe [VmServerAddress])

The server addresses.

getServers_nextToken :: Lens' GetServers (Maybe Text)

The token for the next set of results.

getServers_maxResults :: Lens' GetServers (Maybe Int)

The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned NextToken value.

getServersResponse_serverCatalogStatus :: Lens' GetServersResponse (Maybe ServerCatalogStatus)

The status of the server catalog.

getServersResponse_lastModifiedOn :: Lens' GetServersResponse (Maybe UTCTime)

The time when the server was last modified.

getServersResponse_nextToken :: Lens' GetServersResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null when there are no more results to return.

getServersResponse_serverList :: Lens' GetServersResponse (Maybe [Server])

Information about the servers.

getServersResponse_httpStatus :: Lens' GetServersResponse Int

The response's http status code.

DeleteApp

deleteApp_forceTerminateApp :: Lens' DeleteApp (Maybe Bool)

Indicates whether to terminate the stack corresponding to the application while deleting the application.

deleteApp_appId :: Lens' DeleteApp (Maybe Text)

The ID of the application.

deleteApp_forceStopAppReplication :: Lens' DeleteApp (Maybe Bool)

Indicates whether to stop all replication jobs corresponding to the servers in the application while deleting the application.

deleteAppResponse_httpStatus :: Lens' DeleteAppResponse Int

The response's http status code.

UpdateApp

updateApp_roleName :: Lens' UpdateApp (Maybe Text)

The name of the service role in the customer's account used by AWS SMS.

updateApp_appId :: Lens' UpdateApp (Maybe Text)

The ID of the application.

updateApp_name :: Lens' UpdateApp (Maybe Text)

The new name of the application.

updateApp_description :: Lens' UpdateApp (Maybe Text)

The new description of the application.

updateApp_serverGroups :: Lens' UpdateApp (Maybe [ServerGroup])

The server groups in the application to update.

updateApp_tags :: Lens' UpdateApp (Maybe [Tag])

The tags to associate with the application.

updateAppResponse_appSummary :: Lens' UpdateAppResponse (Maybe AppSummary)

A summary description of the application.

updateAppResponse_serverGroups :: Lens' UpdateAppResponse (Maybe [ServerGroup])

The updated server groups in the application.

updateAppResponse_tags :: Lens' UpdateAppResponse (Maybe [Tag])

The tags associated with the application.

updateAppResponse_httpStatus :: Lens' UpdateAppResponse Int

The response's http status code.

StartOnDemandAppReplication

startOnDemandAppReplication_description :: Lens' StartOnDemandAppReplication (Maybe Text)

The description of the replication run.

startOnDemandAppReplication_appId :: Lens' StartOnDemandAppReplication Text

The ID of the application.

startOnDemandAppReplicationResponse_httpStatus :: Lens' StartOnDemandAppReplicationResponse Int

The response's http status code.

ImportServerCatalog

importServerCatalogResponse_httpStatus :: Lens' ImportServerCatalogResponse Int

The response's http status code.

GenerateTemplate

generateTemplate_appId :: Lens' GenerateTemplate (Maybe Text)

The ID of the application associated with the AWS CloudFormation template.

generateTemplate_templateFormat :: Lens' GenerateTemplate (Maybe OutputFormat)

The format for generating the AWS CloudFormation template.

generateTemplateResponse_s3Location :: Lens' GenerateTemplateResponse (Maybe S3Location)

The location of the Amazon S3 object.

generateTemplateResponse_httpStatus :: Lens' GenerateTemplateResponse Int

The response's http status code.

GetConnectors

getConnectors_nextToken :: Lens' GetConnectors (Maybe Text)

The token for the next set of results.

getConnectors_maxResults :: Lens' GetConnectors (Maybe Int)

The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned NextToken value.

getConnectorsResponse_connectorList :: Lens' GetConnectorsResponse (Maybe [Connector])

Information about the registered connectors.

getConnectorsResponse_nextToken :: Lens' GetConnectorsResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null when there are no more results to return.

getConnectorsResponse_httpStatus :: Lens' GetConnectorsResponse Int

The response's http status code.

GetReplicationJobs

getReplicationJobs_replicationJobId :: Lens' GetReplicationJobs (Maybe Text)

The ID of the replication job.

getReplicationJobs_nextToken :: Lens' GetReplicationJobs (Maybe Text)

The token for the next set of results.

getReplicationJobs_maxResults :: Lens' GetReplicationJobs (Maybe Int)

The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned NextToken value.

getReplicationJobsResponse_replicationJobList :: Lens' GetReplicationJobsResponse (Maybe [ReplicationJob])

Information about the replication jobs.

getReplicationJobsResponse_nextToken :: Lens' GetReplicationJobsResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null when there are no more results to return.

getReplicationJobsResponse_httpStatus :: Lens' GetReplicationJobsResponse Int

The response's http status code.

DisassociateConnector

disassociateConnector_connectorId :: Lens' DisassociateConnector Text

The ID of the connector.

disassociateConnectorResponse_httpStatus :: Lens' DisassociateConnectorResponse Int

The response's http status code.

LaunchApp

launchApp_appId :: Lens' LaunchApp (Maybe Text)

The ID of the application.

launchAppResponse_httpStatus :: Lens' LaunchAppResponse Int

The response's http status code.

GetAppValidationConfiguration

getAppValidationConfiguration_appId :: Lens' GetAppValidationConfiguration Text

The ID of the application.

getAppValidationConfigurationResponse_serverGroupValidationConfigurations :: Lens' GetAppValidationConfigurationResponse (Maybe [ServerGroupValidationConfiguration])

The configuration for instance validation.

getAppValidationConfigurationResponse_appValidationConfigurations :: Lens' GetAppValidationConfigurationResponse (Maybe [AppValidationConfiguration])

The configuration for application validation.

getAppValidationConfigurationResponse_httpStatus :: Lens' GetAppValidationConfigurationResponse Int

The response's http status code.

CreateReplicationJob

createReplicationJob_frequency :: Lens' CreateReplicationJob (Maybe Int)

The time between consecutive replication runs, in hours.

createReplicationJob_numberOfRecentAmisToKeep :: Lens' CreateReplicationJob (Maybe Int)

The maximum number of SMS-created AMIs to retain. The oldest is deleted after the maximum number is reached and a new AMI is created.

createReplicationJob_licenseType :: Lens' CreateReplicationJob (Maybe LicenseType)

The license type to be used for the AMI created by a successful replication run.

createReplicationJob_roleName :: Lens' CreateReplicationJob (Maybe Text)

The name of the IAM role to be used by the AWS SMS.

createReplicationJob_encrypted :: Lens' CreateReplicationJob (Maybe Bool)

Indicates whether the replication job produces encrypted AMIs.

createReplicationJob_kmsKeyId :: Lens' CreateReplicationJob (Maybe Text)

The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is true but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

createReplicationJob_runOnce :: Lens' CreateReplicationJob (Maybe Bool)

Indicates whether to run the replication job one time.

createReplicationJob_description :: Lens' CreateReplicationJob (Maybe Text)

The description of the replication job.

createReplicationJob_serverId :: Lens' CreateReplicationJob Text

The ID of the server.

createReplicationJob_seedReplicationTime :: Lens' CreateReplicationJob UTCTime

The seed replication time.

createReplicationJobResponse_replicationJobId :: Lens' CreateReplicationJobResponse (Maybe Text)

The unique identifier of the replication job.

createReplicationJobResponse_httpStatus :: Lens' CreateReplicationJobResponse Int

The response's http status code.

GenerateChangeSet

generateChangeSet_appId :: Lens' GenerateChangeSet (Maybe Text)

The ID of the application associated with the change set.

generateChangeSet_changesetFormat :: Lens' GenerateChangeSet (Maybe OutputFormat)

The format for the change set.

generateChangeSetResponse_s3Location :: Lens' GenerateChangeSetResponse (Maybe S3Location)

The location of the Amazon S3 object.

generateChangeSetResponse_httpStatus :: Lens' GenerateChangeSetResponse Int

The response's http status code.

GetApp

getApp_appId :: Lens' GetApp (Maybe Text)

The ID of the application.

getAppResponse_appSummary :: Lens' GetAppResponse (Maybe AppSummary)

Information about the application.

getAppResponse_serverGroups :: Lens' GetAppResponse (Maybe [ServerGroup])

The server groups that belong to the application.

getAppResponse_tags :: Lens' GetAppResponse (Maybe [Tag])

The tags associated with the application.

getAppResponse_httpStatus :: Lens' GetAppResponse Int

The response's http status code.

UpdateReplicationJob

updateReplicationJob_frequency :: Lens' UpdateReplicationJob (Maybe Int)

The time between consecutive replication runs, in hours.

updateReplicationJob_numberOfRecentAmisToKeep :: Lens' UpdateReplicationJob (Maybe Int)

The maximum number of SMS-created AMIs to retain. The oldest is deleted after the maximum number is reached and a new AMI is created.

updateReplicationJob_licenseType :: Lens' UpdateReplicationJob (Maybe LicenseType)

The license type to be used for the AMI created by a successful replication run.

updateReplicationJob_roleName :: Lens' UpdateReplicationJob (Maybe Text)

The name of the IAM role to be used by AWS SMS.

updateReplicationJob_encrypted :: Lens' UpdateReplicationJob (Maybe Bool)

When true, the replication job produces encrypted AMIs. For more information, KmsKeyId.

updateReplicationJob_nextReplicationRunStartTime :: Lens' UpdateReplicationJob (Maybe UTCTime)

The start time of the next replication run.

updateReplicationJob_kmsKeyId :: Lens' UpdateReplicationJob (Maybe Text)

The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is enabled but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

updateReplicationJob_description :: Lens' UpdateReplicationJob (Maybe Text)

The description of the replication job.

updateReplicationJob_replicationJobId :: Lens' UpdateReplicationJob Text

The ID of the replication job.

updateReplicationJobResponse_httpStatus :: Lens' UpdateReplicationJobResponse Int

The response's http status code.

DeleteReplicationJob

deleteReplicationJob_replicationJobId :: Lens' DeleteReplicationJob Text

The ID of the replication job.

deleteReplicationJobResponse_httpStatus :: Lens' DeleteReplicationJobResponse Int

The response's http status code.

CreateApp

createApp_clientToken :: Lens' CreateApp (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of application creation.

createApp_roleName :: Lens' CreateApp (Maybe Text)

The name of the service role in the customer's account to be used by AWS SMS.

createApp_name :: Lens' CreateApp (Maybe Text)

The name of the new application.

createApp_description :: Lens' CreateApp (Maybe Text)

The description of the new application

createApp_serverGroups :: Lens' CreateApp (Maybe [ServerGroup])

The server groups to include in the application.

createApp_tags :: Lens' CreateApp (Maybe [Tag])

The tags to be associated with the application.

createAppResponse_appSummary :: Lens' CreateAppResponse (Maybe AppSummary)

A summary description of the application.

createAppResponse_serverGroups :: Lens' CreateAppResponse (Maybe [ServerGroup])

The server groups included in the application.

createAppResponse_tags :: Lens' CreateAppResponse (Maybe [Tag])

The tags associated with the application.

createAppResponse_httpStatus :: Lens' CreateAppResponse Int

The response's http status code.

StopAppReplication

stopAppReplication_appId :: Lens' StopAppReplication (Maybe Text)

The ID of the application.

stopAppReplicationResponse_httpStatus :: Lens' StopAppReplicationResponse Int

The response's http status code.

DeleteAppValidationConfiguration

deleteAppValidationConfiguration_appId :: Lens' DeleteAppValidationConfiguration Text

The ID of the application.

deleteAppValidationConfigurationResponse_httpStatus :: Lens' DeleteAppValidationConfigurationResponse Int

The response's http status code.

PutAppValidationConfiguration

putAppValidationConfiguration_serverGroupValidationConfigurations :: Lens' PutAppValidationConfiguration (Maybe [ServerGroupValidationConfiguration])

The configuration for instance validation.

putAppValidationConfiguration_appValidationConfigurations :: Lens' PutAppValidationConfiguration (Maybe [AppValidationConfiguration])

The configuration for application validation.

putAppValidationConfiguration_appId :: Lens' PutAppValidationConfiguration Text

The ID of the application.

putAppValidationConfigurationResponse_httpStatus :: Lens' PutAppValidationConfigurationResponse Int

The response's http status code.

GetAppValidationOutput

getAppValidationOutput_appId :: Lens' GetAppValidationOutput Text

The ID of the application.

getAppValidationOutputResponse_validationOutputList :: Lens' GetAppValidationOutputResponse (Maybe [ValidationOutput])

The validation output.

getAppValidationOutputResponse_httpStatus :: Lens' GetAppValidationOutputResponse Int

The response's http status code.

GetAppReplicationConfiguration

getAppReplicationConfiguration_appId :: Lens' GetAppReplicationConfiguration (Maybe Text)

The ID of the application.

getAppReplicationConfigurationResponse_serverGroupReplicationConfigurations :: Lens' GetAppReplicationConfigurationResponse (Maybe [ServerGroupReplicationConfiguration])

The replication configurations associated with server groups in this application.

getAppReplicationConfigurationResponse_httpStatus :: Lens' GetAppReplicationConfigurationResponse Int

The response's http status code.

StartOnDemandReplicationRun

startOnDemandReplicationRun_description :: Lens' StartOnDemandReplicationRun (Maybe Text)

The description of the replication run.

startOnDemandReplicationRun_replicationJobId :: Lens' StartOnDemandReplicationRun Text

The ID of the replication job.

startOnDemandReplicationRunResponse_replicationRunId :: Lens' StartOnDemandReplicationRunResponse (Maybe Text)

The ID of the replication run.

startOnDemandReplicationRunResponse_httpStatus :: Lens' StartOnDemandReplicationRunResponse Int

The response's http status code.

NotifyAppValidationOutput

notifyAppValidationOutput_notificationContext :: Lens' NotifyAppValidationOutput (Maybe NotificationContext)

The notification information.

notifyAppValidationOutput_appId :: Lens' NotifyAppValidationOutput Text

The ID of the application.

notifyAppValidationOutputResponse_httpStatus :: Lens' NotifyAppValidationOutputResponse Int

The response's http status code.

Types

AppSummary

appSummary_creationTime :: Lens' AppSummary (Maybe UTCTime)

The creation time of the application.

appSummary_totalServers :: Lens' AppSummary (Maybe Int)

The number of servers present in the application.

appSummary_status :: Lens' AppSummary (Maybe AppStatus)

Status of the application.

appSummary_launchDetails :: Lens' AppSummary (Maybe LaunchDetails)

Details about the latest launch of the application.

appSummary_launchStatusMessage :: Lens' AppSummary (Maybe Text)

A message related to the launch status of the application.

appSummary_replicationConfigurationStatus :: Lens' AppSummary (Maybe AppReplicationConfigurationStatus)

Status of the replication configuration.

appSummary_replicationStatusMessage :: Lens' AppSummary (Maybe Text)

A message related to the replication status of the application.

appSummary_totalServerGroups :: Lens' AppSummary (Maybe Int)

The number of server groups present in the application.

appSummary_roleName :: Lens' AppSummary (Maybe Text)

The name of the service role in the customer's account used by AWS SMS.

appSummary_launchConfigurationStatus :: Lens' AppSummary (Maybe AppLaunchConfigurationStatus)

Status of the launch configuration.

appSummary_launchStatus :: Lens' AppSummary (Maybe AppLaunchStatus)

The launch status of the application.

appSummary_appId :: Lens' AppSummary (Maybe Text)

The unique ID of the application.

appSummary_name :: Lens' AppSummary (Maybe Text)

The name of the application.

appSummary_statusMessage :: Lens' AppSummary (Maybe Text)

A message related to the status of the application

appSummary_latestReplicationTime :: Lens' AppSummary (Maybe UTCTime)

The timestamp of the application's most recent successful replication.

appSummary_importedAppId :: Lens' AppSummary (Maybe Text)

The ID of the application.

appSummary_replicationStatus :: Lens' AppSummary (Maybe AppReplicationStatus)

The replication status of the application.

appSummary_lastModified :: Lens' AppSummary (Maybe UTCTime)

The last modified time of the application.

appSummary_description :: Lens' AppSummary (Maybe Text)

The description of the application.

AppValidationConfiguration

appValidationConfiguration_ssmValidationParameters :: Lens' AppValidationConfiguration (Maybe SSMValidationParameters)

The validation parameters.

appValidationConfiguration_name :: Lens' AppValidationConfiguration (Maybe Text)

The name of the configuration.

appValidationConfiguration_validationId :: Lens' AppValidationConfiguration (Maybe Text)

The ID of the validation.

appValidationConfiguration_appValidationStrategy :: Lens' AppValidationConfiguration (Maybe AppValidationStrategy)

The validation strategy.

AppValidationOutput

appValidationOutput_ssmOutput :: Lens' AppValidationOutput (Maybe SSMOutput)

Output from using SSM to validate the application.

Connector

connector_status :: Lens' Connector (Maybe ConnectorStatus)

The status of the connector.

connector_vmManagerName :: Lens' Connector (Maybe Text)

The name of the VM manager.

connector_ipAddress :: Lens' Connector (Maybe Text)

The IP address of the connector.

connector_vmManagerId :: Lens' Connector (Maybe Text)

The ID of the VM manager.

connector_vmManagerType :: Lens' Connector (Maybe VmManagerType)

The VM management product.

connector_connectorId :: Lens' Connector (Maybe Text)

The ID of the connector.

connector_associatedOn :: Lens' Connector (Maybe UTCTime)

The time the connector was associated.

connector_macAddress :: Lens' Connector (Maybe Text)

The MAC address of the connector.

connector_version :: Lens' Connector (Maybe Text)

The connector version.

connector_capabilityList :: Lens' Connector (Maybe [ConnectorCapability])

The capabilities of the connector.

LaunchDetails

launchDetails_stackId :: Lens' LaunchDetails (Maybe Text)

The ID of the latest stack launched for this application.

launchDetails_latestLaunchTime :: Lens' LaunchDetails (Maybe UTCTime)

The latest time that this application was launched successfully.

launchDetails_stackName :: Lens' LaunchDetails (Maybe Text)

The name of the latest stack launched for this application.

NotificationContext

notificationContext_status :: Lens' NotificationContext (Maybe ValidationStatus)

The status of the validation.

notificationContext_statusMessage :: Lens' NotificationContext (Maybe Text)

The status message.

notificationContext_validationId :: Lens' NotificationContext (Maybe Text)

The ID of the validation.

ReplicationJob

replicationJob_frequency :: Lens' ReplicationJob (Maybe Int)

The time between consecutive replication runs, in hours.

replicationJob_numberOfRecentAmisToKeep :: Lens' ReplicationJob (Maybe Int)

The number of recent AMIs to keep in the customer's account for a replication job. By default, the value is set to zero, meaning that all AMIs are kept.

replicationJob_state :: Lens' ReplicationJob (Maybe ReplicationJobState)

The state of the replication job.

replicationJob_serverType :: Lens' ReplicationJob (Maybe ServerType)

The type of server.

replicationJob_serverId :: Lens' ReplicationJob (Maybe Text)

The ID of the server.

replicationJob_licenseType :: Lens' ReplicationJob (Maybe LicenseType)

The license type to be used for the AMI created by a successful replication run.

replicationJob_roleName :: Lens' ReplicationJob (Maybe Text)

The name of the IAM role to be used by AWS SMS.

replicationJob_vmServer :: Lens' ReplicationJob (Maybe VmServer)

Information about the VM server.

replicationJob_encrypted :: Lens' ReplicationJob (Maybe Bool)

Indicates whether the replication job should produce encrypted AMIs.

replicationJob_replicationJobId :: Lens' ReplicationJob (Maybe Text)

The ID of the replication job.

replicationJob_replicationRunList :: Lens' ReplicationJob (Maybe [ReplicationRun])

Information about the replication runs.

replicationJob_nextReplicationRunStartTime :: Lens' ReplicationJob (Maybe UTCTime)

The start time of the next replication run.

replicationJob_statusMessage :: Lens' ReplicationJob (Maybe Text)

The description of the current status of the replication job.

replicationJob_kmsKeyId :: Lens' ReplicationJob (Maybe Text)

The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is enabled but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

replicationJob_latestAmiId :: Lens' ReplicationJob (Maybe Text)

The ID of the latest Amazon Machine Image (AMI).

replicationJob_seedReplicationTime :: Lens' ReplicationJob (Maybe UTCTime)

The seed replication time.

replicationJob_runOnce :: Lens' ReplicationJob (Maybe Bool)

Indicates whether to run the replication job one time.

replicationJob_description :: Lens' ReplicationJob (Maybe Text)

The description of the replication job.

ReplicationRun

replicationRun_state :: Lens' ReplicationRun (Maybe ReplicationRunState)

The state of the replication run.

replicationRun_replicationRunId :: Lens' ReplicationRun (Maybe Text)

The ID of the replication run.

replicationRun_encrypted :: Lens' ReplicationRun (Maybe Bool)

Indicates whether the replication run should produce an encrypted AMI.

replicationRun_stageDetails :: Lens' ReplicationRun (Maybe ReplicationRunStageDetails)

Details about the current stage of the replication run.

replicationRun_scheduledStartTime :: Lens' ReplicationRun (Maybe UTCTime)

The start time of the next replication run.

replicationRun_statusMessage :: Lens' ReplicationRun (Maybe Text)

The description of the current status of the replication job.

replicationRun_kmsKeyId :: Lens' ReplicationRun (Maybe Text)

The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is true but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

replicationRun_completedTime :: Lens' ReplicationRun (Maybe UTCTime)

The completion time of the last replication run.

replicationRun_amiId :: Lens' ReplicationRun (Maybe Text)

The ID of the Amazon Machine Image (AMI) from the replication run.

replicationRun_type :: Lens' ReplicationRun (Maybe ReplicationRunType)

The type of replication run.

replicationRun_description :: Lens' ReplicationRun (Maybe Text)

The description of the replication run.

ReplicationRunStageDetails

replicationRunStageDetails_stage :: Lens' ReplicationRunStageDetails (Maybe Text)

The current stage of a replication run.

replicationRunStageDetails_stageProgress :: Lens' ReplicationRunStageDetails (Maybe Text)

The progress of the current stage of a replication run.

S3Location

s3Location_bucket :: Lens' S3Location (Maybe Text)

The Amazon S3 bucket name.

s3Location_key :: Lens' S3Location (Maybe Text)

The Amazon S3 bucket key.

SSMOutput

sSMOutput_s3Location :: Lens' SSMOutput (Maybe S3Location)

Undocumented member.

SSMValidationParameters

sSMValidationParameters_instanceId :: Lens' SSMValidationParameters (Maybe Text)

The ID of the instance. The instance must have the following tag: UserForSMSApplicationValidation=true.

sSMValidationParameters_command :: Lens' SSMValidationParameters (Maybe Text)

The command to run the validation script

sSMValidationParameters_executionTimeoutSeconds :: Lens' SSMValidationParameters (Maybe Natural)

The timeout interval, in seconds.

sSMValidationParameters_scriptType :: Lens' SSMValidationParameters (Maybe ScriptType)

The type of validation script.

sSMValidationParameters_source :: Lens' SSMValidationParameters (Maybe Source)

The location of the validation script.

sSMValidationParameters_outputS3BucketName :: Lens' SSMValidationParameters (Maybe Text)

The name of the S3 bucket for output.

Server

server_serverType :: Lens' Server (Maybe ServerType)

The type of server.

server_serverId :: Lens' Server (Maybe Text)

The ID of the server.

server_replicationJobTerminated :: Lens' Server (Maybe Bool)

Indicates whether the replication job is deleted or failed.

server_vmServer :: Lens' Server (Maybe VmServer)

Information about the VM server.

server_replicationJobId :: Lens' Server (Maybe Text)

The ID of the replication job.

ServerGroup

serverGroup_serverList :: Lens' ServerGroup (Maybe [Server])

The servers that belong to a server group.

serverGroup_name :: Lens' ServerGroup (Maybe Text)

The name of a server group.

serverGroup_serverGroupId :: Lens' ServerGroup (Maybe Text)

The ID of a server group.

ServerGroupLaunchConfiguration

serverGroupLaunchConfiguration_serverGroupId :: Lens' ServerGroupLaunchConfiguration (Maybe Text)

The ID of the server group with which the launch configuration is associated.

serverGroupLaunchConfiguration_launchOrder :: Lens' ServerGroupLaunchConfiguration (Maybe Int)

The launch order of servers in the server group.

serverGroupLaunchConfiguration_serverLaunchConfigurations :: Lens' ServerGroupLaunchConfiguration (Maybe [ServerLaunchConfiguration])

The launch configuration for servers in the server group.

ServerGroupReplicationConfiguration

serverGroupReplicationConfiguration_serverGroupId :: Lens' ServerGroupReplicationConfiguration (Maybe Text)

The ID of the server group with which this replication configuration is associated.

serverGroupReplicationConfiguration_serverReplicationConfigurations :: Lens' ServerGroupReplicationConfiguration (Maybe [ServerReplicationConfiguration])

The replication configuration for servers in the server group.

ServerGroupValidationConfiguration

serverGroupValidationConfiguration_serverValidationConfigurations :: Lens' ServerGroupValidationConfiguration (Maybe [ServerValidationConfiguration])

The validation configuration.

serverGroupValidationConfiguration_serverGroupId :: Lens' ServerGroupValidationConfiguration (Maybe Text)

The ID of the server group.

ServerLaunchConfiguration

serverLaunchConfiguration_ec2KeyName :: Lens' ServerLaunchConfiguration (Maybe Text)

The name of the Amazon EC2 SSH key to be used for connecting to the launched server.

serverLaunchConfiguration_configureScriptType :: Lens' ServerLaunchConfiguration (Maybe ScriptType)

The type of configuration script.

serverLaunchConfiguration_associatePublicIpAddress :: Lens' ServerLaunchConfiguration (Maybe Bool)

Indicates whether a publicly accessible IP address is created when launching the server.

serverLaunchConfiguration_iamInstanceProfileName :: Lens' ServerLaunchConfiguration (Maybe Text)

The name of the IAM instance profile.

serverLaunchConfiguration_subnet :: Lens' ServerLaunchConfiguration (Maybe Text)

The ID of the subnet the server should be launched into.

serverLaunchConfiguration_logicalId :: Lens' ServerLaunchConfiguration (Maybe Text)

The logical ID of the server in the AWS CloudFormation template.

serverLaunchConfiguration_securityGroup :: Lens' ServerLaunchConfiguration (Maybe Text)

The ID of the security group that applies to the launched server.

serverLaunchConfiguration_userData :: Lens' ServerLaunchConfiguration (Maybe UserData)

Location of the user-data script to be executed when launching the server.

serverLaunchConfiguration_instanceType :: Lens' ServerLaunchConfiguration (Maybe Text)

The instance type to use when launching the server.

serverLaunchConfiguration_configureScript :: Lens' ServerLaunchConfiguration (Maybe S3Location)

Undocumented member.

serverLaunchConfiguration_server :: Lens' ServerLaunchConfiguration (Maybe Server)

The ID of the server with which the launch configuration is associated.

serverLaunchConfiguration_vpc :: Lens' ServerLaunchConfiguration (Maybe Text)

The ID of the VPC into which the server should be launched.

ServerReplicationConfiguration

serverReplicationConfiguration_serverReplicationParameters :: Lens' ServerReplicationConfiguration (Maybe ServerReplicationParameters)

The parameters for replicating the server.

serverReplicationConfiguration_server :: Lens' ServerReplicationConfiguration (Maybe Server)

The ID of the server with which this replication configuration is associated.

ServerReplicationParameters

serverReplicationParameters_frequency :: Lens' ServerReplicationParameters (Maybe Int)

The frequency of creating replication jobs for the server.

serverReplicationParameters_numberOfRecentAmisToKeep :: Lens' ServerReplicationParameters (Maybe Int)

The number of recent AMIs to keep when creating a replication job for this server.

serverReplicationParameters_seedTime :: Lens' ServerReplicationParameters (Maybe UTCTime)

The seed time for creating a replication job for the server.

serverReplicationParameters_licenseType :: Lens' ServerReplicationParameters (Maybe LicenseType)

The license type for creating a replication job for the server.

serverReplicationParameters_encrypted :: Lens' ServerReplicationParameters (Maybe Bool)

Indicates whether the replication job produces encrypted AMIs.

serverReplicationParameters_kmsKeyId :: Lens' ServerReplicationParameters (Maybe Text)

The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is enabled but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

serverReplicationParameters_runOnce :: Lens' ServerReplicationParameters (Maybe Bool)

Indicates whether to run the replication job one time.

ServerValidationConfiguration

serverValidationConfiguration_serverValidationStrategy :: Lens' ServerValidationConfiguration (Maybe ServerValidationStrategy)

The validation strategy.

serverValidationConfiguration_userDataValidationParameters :: Lens' ServerValidationConfiguration (Maybe UserDataValidationParameters)

The validation parameters.

serverValidationConfiguration_name :: Lens' ServerValidationConfiguration (Maybe Text)

The name of the configuration.

serverValidationConfiguration_server :: Lens' ServerValidationConfiguration (Maybe Server)

Undocumented member.

serverValidationConfiguration_validationId :: Lens' ServerValidationConfiguration (Maybe Text)

The ID of the validation.

ServerValidationOutput

serverValidationOutput_server :: Lens' ServerValidationOutput (Maybe Server)

Undocumented member.

Source

source_s3Location :: Lens' Source (Maybe S3Location)

Undocumented member.

Tag

tag_value :: Lens' Tag (Maybe Text)

The tag value.

tag_key :: Lens' Tag (Maybe Text)

The tag key.

UserData

userData_s3Location :: Lens' UserData (Maybe S3Location)

Amazon S3 location of the user-data script.

UserDataValidationParameters

userDataValidationParameters_scriptType :: Lens' UserDataValidationParameters (Maybe ScriptType)

The type of validation script.

userDataValidationParameters_source :: Lens' UserDataValidationParameters (Maybe Source)

The location of the validation script.

ValidationOutput

validationOutput_status :: Lens' ValidationOutput (Maybe ValidationStatus)

The status of the validation.

validationOutput_appValidationOutput :: Lens' ValidationOutput (Maybe AppValidationOutput)

The output from validating an application.

validationOutput_latestValidationTime :: Lens' ValidationOutput (Maybe UTCTime)

The latest time that the validation was performed.

validationOutput_name :: Lens' ValidationOutput (Maybe Text)

The name of the validation.

validationOutput_statusMessage :: Lens' ValidationOutput (Maybe Text)

The status message.

validationOutput_validationId :: Lens' ValidationOutput (Maybe Text)

The ID of the validation.

validationOutput_serverValidationOutput :: Lens' ValidationOutput (Maybe ServerValidationOutput)

The output from validation an instance.

VmServer

vmServer_vmManagerName :: Lens' VmServer (Maybe Text)

The name of the VM manager.

vmServer_vmManagerType :: Lens' VmServer (Maybe VmManagerType)

The type of VM management product.

vmServer_vmServerAddress :: Lens' VmServer (Maybe VmServerAddress)

The VM server location.

vmServer_vmName :: Lens' VmServer (Maybe Text)

The name of the VM.

vmServer_vmPath :: Lens' VmServer (Maybe Text)

The VM folder path in the vCenter Server virtual machine inventory tree.

VmServerAddress

vmServerAddress_vmManagerId :: Lens' VmServerAddress (Maybe Text)

The ID of the VM manager.

vmServerAddress_vmId :: Lens' VmServerAddress (Maybe Text)

The ID of the VM.