Network.AWS.SMS.PutAppLaunchConfiguration

Description

Creates or updates the launch configuration for the specified application.

Creating a Request

data PutAppLaunchConfiguration

See: newPutAppLaunchConfiguration smart constructor.

Constructors

PutAppLaunchConfiguration'

Fields

appId :: Maybe Text

The ID of the application.

roleName :: Maybe Text

The name of service role in the customer's account that AWS CloudFormation uses to launch the application.

autoLaunch :: Maybe Bool

Indicates whether the application is configured to launch automatically after replication is complete.

serverGroupLaunchConfigurations :: Maybe [ServerGroupLaunchConfiguration]

Information about the launch configurations for server groups in the application.

newPutAppLaunchConfiguration :: PutAppLaunchConfiguration

Create a value of PutAppLaunchConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appId:PutAppLaunchConfiguration', putAppLaunchConfiguration_appId - The ID of the application.

$sel:roleName:PutAppLaunchConfiguration', putAppLaunchConfiguration_roleName - The name of service role in the customer's account that AWS CloudFormation uses to launch the application.

$sel:autoLaunch:PutAppLaunchConfiguration', putAppLaunchConfiguration_autoLaunch - Indicates whether the application is configured to launch automatically after replication is complete.

$sel:serverGroupLaunchConfigurations:PutAppLaunchConfiguration', putAppLaunchConfiguration_serverGroupLaunchConfigurations - Information about the launch configurations for server groups in the application.

Request Lenses

putAppLaunchConfiguration_appId :: Lens' PutAppLaunchConfiguration (Maybe Text)

The ID of the application.

putAppLaunchConfiguration_roleName :: Lens' PutAppLaunchConfiguration (Maybe Text)

The name of service role in the customer's account that AWS CloudFormation uses to launch the application.

putAppLaunchConfiguration_autoLaunch :: Lens' PutAppLaunchConfiguration (Maybe Bool)

Indicates whether the application is configured to launch automatically after replication is complete.

putAppLaunchConfiguration_serverGroupLaunchConfigurations :: Lens' PutAppLaunchConfiguration (Maybe [ServerGroupLaunchConfiguration])

Information about the launch configurations for server groups in the application.

Destructuring the Response

data PutAppLaunchConfigurationResponse

See: newPutAppLaunchConfigurationResponse smart constructor.

Constructors

PutAppLaunchConfigurationResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutAppLaunchConfigurationResponse

Arguments

:: Int
$sel:httpStatus:PutAppLaunchConfigurationResponse'

-> PutAppLaunchConfigurationResponse

Create a value of PutAppLaunchConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutAppLaunchConfigurationResponse', putAppLaunchConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

putAppLaunchConfigurationResponse_httpStatus :: Lens' PutAppLaunchConfigurationResponse Int

The response's http status code.