Network.AWS.SMS.Types.AppSummary

Description

Documentation

data AppSummary

Information about the application.

See: newAppSummary smart constructor.

Constructors

AppSummary'

Fields

creationTime :: Maybe POSIX

The creation time of the application.

totalServers :: Maybe Int

The number of servers present in the application.

status :: Maybe AppStatus

Status of the application.

launchDetails :: Maybe LaunchDetails

Details about the latest launch of the application.

launchStatusMessage :: Maybe Text

A message related to the launch status of the application.

replicationConfigurationStatus :: Maybe AppReplicationConfigurationStatus

Status of the replication configuration.

replicationStatusMessage :: Maybe Text

A message related to the replication status of the application.

totalServerGroups :: Maybe Int

The number of server groups present in the application.

roleName :: Maybe Text

The name of the service role in the customer's account used by AWS SMS.

launchConfigurationStatus :: Maybe AppLaunchConfigurationStatus

Status of the launch configuration.

launchStatus :: Maybe AppLaunchStatus

The launch status of the application.

appId :: Maybe Text

The unique ID of the application.

name :: Maybe Text

The name of the application.

statusMessage :: Maybe Text

A message related to the status of the application

latestReplicationTime :: Maybe POSIX

The timestamp of the application's most recent successful replication.

importedAppId :: Maybe Text

The ID of the application.

replicationStatus :: Maybe AppReplicationStatus

The replication status of the application.

lastModified :: Maybe POSIX

The last modified time of the application.

description :: Maybe Text

The description of the application.

newAppSummary :: AppSummary

Create a value of AppSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:AppSummary', appSummary_creationTime - The creation time of the application.

$sel:totalServers:AppSummary', appSummary_totalServers - The number of servers present in the application.

$sel:status:AppSummary', appSummary_status - Status of the application.

$sel:launchDetails:AppSummary', appSummary_launchDetails - Details about the latest launch of the application.

$sel:launchStatusMessage:AppSummary', appSummary_launchStatusMessage - A message related to the launch status of the application.

$sel:replicationConfigurationStatus:AppSummary', appSummary_replicationConfigurationStatus - Status of the replication configuration.

$sel:replicationStatusMessage:AppSummary', appSummary_replicationStatusMessage - A message related to the replication status of the application.

$sel:totalServerGroups:AppSummary', appSummary_totalServerGroups - The number of server groups present in the application.

$sel:roleName:AppSummary', appSummary_roleName - The name of the service role in the customer's account used by AWS SMS.

$sel:launchConfigurationStatus:AppSummary', appSummary_launchConfigurationStatus - Status of the launch configuration.

$sel:launchStatus:AppSummary', appSummary_launchStatus - The launch status of the application.

$sel:appId:AppSummary', appSummary_appId - The unique ID of the application.

$sel:name:AppSummary', appSummary_name - The name of the application.

$sel:statusMessage:AppSummary', appSummary_statusMessage - A message related to the status of the application

$sel:latestReplicationTime:AppSummary', appSummary_latestReplicationTime - The timestamp of the application's most recent successful replication.

$sel:importedAppId:AppSummary', appSummary_importedAppId - The ID of the application.

$sel:replicationStatus:AppSummary', appSummary_replicationStatus - The replication status of the application.

$sel:lastModified:AppSummary', appSummary_lastModified - The last modified time of the application.

$sel:description:AppSummary', appSummary_description - The description of the application.

appSummary_creationTime :: Lens' AppSummary (Maybe UTCTime)

The creation time of the application.

appSummary_totalServers :: Lens' AppSummary (Maybe Int)

The number of servers present in the application.

appSummary_status :: Lens' AppSummary (Maybe AppStatus)

Status of the application.

appSummary_launchDetails :: Lens' AppSummary (Maybe LaunchDetails)

Details about the latest launch of the application.

appSummary_launchStatusMessage :: Lens' AppSummary (Maybe Text)

A message related to the launch status of the application.

appSummary_replicationConfigurationStatus :: Lens' AppSummary (Maybe AppReplicationConfigurationStatus)

Status of the replication configuration.

appSummary_replicationStatusMessage :: Lens' AppSummary (Maybe Text)

A message related to the replication status of the application.

appSummary_totalServerGroups :: Lens' AppSummary (Maybe Int)

The number of server groups present in the application.

appSummary_roleName :: Lens' AppSummary (Maybe Text)

The name of the service role in the customer's account used by AWS SMS.

appSummary_launchConfigurationStatus :: Lens' AppSummary (Maybe AppLaunchConfigurationStatus)

Status of the launch configuration.

appSummary_launchStatus :: Lens' AppSummary (Maybe AppLaunchStatus)

The launch status of the application.

appSummary_appId :: Lens' AppSummary (Maybe Text)

The unique ID of the application.

appSummary_name :: Lens' AppSummary (Maybe Text)

The name of the application.

appSummary_statusMessage :: Lens' AppSummary (Maybe Text)

A message related to the status of the application

appSummary_latestReplicationTime :: Lens' AppSummary (Maybe UTCTime)

The timestamp of the application's most recent successful replication.

appSummary_importedAppId :: Lens' AppSummary (Maybe Text)

The ID of the application.

appSummary_replicationStatus :: Lens' AppSummary (Maybe AppReplicationStatus)

The replication status of the application.

appSummary_lastModified :: Lens' AppSummary (Maybe UTCTime)

The last modified time of the application.

appSummary_description :: Lens' AppSummary (Maybe Text)

The description of the application.