Network.AWS.SMS.Types.ReplicationJob

Description

Documentation

data ReplicationJob

Represents a replication job.

See: newReplicationJob smart constructor.

Constructors

ReplicationJob'

Fields

nextReplicationRunStartTime :: Maybe POSIX

The start time of the next replication run.

numberOfRecentAmisToKeep :: Maybe Int

The number of recent AMIs to keep in the customer's account for a replication job. By default, the value is set to zero, meaning that all AMIs are kept.

statusMessage :: Maybe Text

The description of the current status of the replication job.

encrypted :: Maybe Bool

Indicates whether the replication job should produce encrypted AMIs.

latestAmiId :: Maybe Text

The ID of the latest Amazon Machine Image (AMI).

roleName :: Maybe Text

The name of the IAM role to be used by AWS SMS.

serverId :: Maybe Text

The ID of the server.

state :: Maybe ReplicationJobState

The state of the replication job.

kmsKeyId :: Maybe Text

The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is enabled but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

replicationRunList :: Maybe [ReplicationRun]

Information about the replication runs.

frequency :: Maybe Int

The time between consecutive replication runs, in hours.

replicationJobId :: Maybe Text

The ID of the replication job.

runOnce :: Maybe Bool

Indicates whether to run the replication job one time.

seedReplicationTime :: Maybe POSIX

The seed replication time.

description :: Maybe Text

The description of the replication job.

vmServer :: Maybe VmServer

Information about the VM server.

licenseType :: Maybe LicenseType

The license type to be used for the AMI created by a successful replication run.

serverType :: Maybe ServerType

The type of server.

newReplicationJob :: ReplicationJob

Create a value of ReplicationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextReplicationRunStartTime:ReplicationJob', replicationJob_nextReplicationRunStartTime - The start time of the next replication run.

$sel:numberOfRecentAmisToKeep:ReplicationJob', replicationJob_numberOfRecentAmisToKeep - The number of recent AMIs to keep in the customer's account for a replication job. By default, the value is set to zero, meaning that all AMIs are kept.

$sel:statusMessage:ReplicationJob', replicationJob_statusMessage - The description of the current status of the replication job.

$sel:encrypted:ReplicationJob', replicationJob_encrypted - Indicates whether the replication job should produce encrypted AMIs.

$sel:latestAmiId:ReplicationJob', replicationJob_latestAmiId - The ID of the latest Amazon Machine Image (AMI).

$sel:roleName:ReplicationJob', replicationJob_roleName - The name of the IAM role to be used by AWS SMS.

$sel:serverId:ReplicationJob', replicationJob_serverId - The ID of the server.

$sel:state:ReplicationJob', replicationJob_state - The state of the replication job.

$sel:kmsKeyId:ReplicationJob', replicationJob_kmsKeyId - The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is enabled but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

$sel:replicationRunList:ReplicationJob', replicationJob_replicationRunList - Information about the replication runs.

$sel:frequency:ReplicationJob', replicationJob_frequency - The time between consecutive replication runs, in hours.

$sel:replicationJobId:ReplicationJob', replicationJob_replicationJobId - The ID of the replication job.

$sel:runOnce:ReplicationJob', replicationJob_runOnce - Indicates whether to run the replication job one time.

$sel:seedReplicationTime:ReplicationJob', replicationJob_seedReplicationTime - The seed replication time.

$sel:description:ReplicationJob', replicationJob_description - The description of the replication job.

$sel:vmServer:ReplicationJob', replicationJob_vmServer - Information about the VM server.

$sel:licenseType:ReplicationJob', replicationJob_licenseType - The license type to be used for the AMI created by a successful replication run.

$sel:serverType:ReplicationJob', replicationJob_serverType - The type of server.

replicationJob_nextReplicationRunStartTime :: Lens' ReplicationJob (Maybe UTCTime)

The start time of the next replication run.

replicationJob_numberOfRecentAmisToKeep :: Lens' ReplicationJob (Maybe Int)

The number of recent AMIs to keep in the customer's account for a replication job. By default, the value is set to zero, meaning that all AMIs are kept.

replicationJob_statusMessage :: Lens' ReplicationJob (Maybe Text)

The description of the current status of the replication job.

replicationJob_encrypted :: Lens' ReplicationJob (Maybe Bool)

Indicates whether the replication job should produce encrypted AMIs.

replicationJob_latestAmiId :: Lens' ReplicationJob (Maybe Text)

The ID of the latest Amazon Machine Image (AMI).

replicationJob_roleName :: Lens' ReplicationJob (Maybe Text)

The name of the IAM role to be used by AWS SMS.

replicationJob_serverId :: Lens' ReplicationJob (Maybe Text)

The ID of the server.

replicationJob_state :: Lens' ReplicationJob (Maybe ReplicationJobState)

The state of the replication job.

replicationJob_kmsKeyId :: Lens' ReplicationJob (Maybe Text)

The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is enabled but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

replicationJob_replicationRunList :: Lens' ReplicationJob (Maybe [ReplicationRun])

Information about the replication runs.

replicationJob_frequency :: Lens' ReplicationJob (Maybe Int)

The time between consecutive replication runs, in hours.

replicationJob_replicationJobId :: Lens' ReplicationJob (Maybe Text)

The ID of the replication job.

replicationJob_runOnce :: Lens' ReplicationJob (Maybe Bool)

Indicates whether to run the replication job one time.

replicationJob_seedReplicationTime :: Lens' ReplicationJob (Maybe UTCTime)

The seed replication time.

replicationJob_description :: Lens' ReplicationJob (Maybe Text)

The description of the replication job.

replicationJob_vmServer :: Lens' ReplicationJob (Maybe VmServer)

Information about the VM server.

replicationJob_licenseType :: Lens' ReplicationJob (Maybe LicenseType)

The license type to be used for the AMI created by a successful replication run.

replicationJob_serverType :: Lens' ReplicationJob (Maybe ServerType)

The type of server.