Network.AWS.SMS.Types.ReplicationRun

Description

Documentation

data ReplicationRun

Represents a replication run.

See: newReplicationRun smart constructor.

Constructors

ReplicationRun'

Fields

statusMessage :: Maybe Text

The description of the current status of the replication job.

encrypted :: Maybe Bool

Indicates whether the replication run should produce an encrypted AMI.

replicationRunId :: Maybe Text

The ID of the replication run.

completedTime :: Maybe POSIX

The completion time of the last replication run.

amiId :: Maybe Text

The ID of the Amazon Machine Image (AMI) from the replication run.

state :: Maybe ReplicationRunState

The state of the replication run.

kmsKeyId :: Maybe Text

The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is true but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

scheduledStartTime :: Maybe POSIX

The start time of the next replication run.

stageDetails :: Maybe ReplicationRunStageDetails

Details about the current stage of the replication run.

description :: Maybe Text

The description of the replication run.

type' :: Maybe ReplicationRunType

The type of replication run.

newReplicationRun :: ReplicationRun

Create a value of ReplicationRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusMessage:ReplicationRun', replicationRun_statusMessage - The description of the current status of the replication job.

$sel:encrypted:ReplicationRun', replicationRun_encrypted - Indicates whether the replication run should produce an encrypted AMI.

$sel:replicationRunId:ReplicationRun', replicationRun_replicationRunId - The ID of the replication run.

$sel:completedTime:ReplicationRun', replicationRun_completedTime - The completion time of the last replication run.

$sel:amiId:ReplicationRun', replicationRun_amiId - The ID of the Amazon Machine Image (AMI) from the replication run.

$sel:state:ReplicationRun', replicationRun_state - The state of the replication run.

$sel:kmsKeyId:ReplicationRun', replicationRun_kmsKeyId - The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is true but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

$sel:scheduledStartTime:ReplicationRun', replicationRun_scheduledStartTime - The start time of the next replication run.

$sel:stageDetails:ReplicationRun', replicationRun_stageDetails - Details about the current stage of the replication run.

$sel:description:ReplicationRun', replicationRun_description - The description of the replication run.

$sel:type':ReplicationRun', replicationRun_type - The type of replication run.

replicationRun_statusMessage :: Lens' ReplicationRun (Maybe Text)

The description of the current status of the replication job.

replicationRun_encrypted :: Lens' ReplicationRun (Maybe Bool)

Indicates whether the replication run should produce an encrypted AMI.

replicationRun_replicationRunId :: Lens' ReplicationRun (Maybe Text)

The ID of the replication run.

replicationRun_completedTime :: Lens' ReplicationRun (Maybe UTCTime)

The completion time of the last replication run.

replicationRun_amiId :: Lens' ReplicationRun (Maybe Text)

The ID of the Amazon Machine Image (AMI) from the replication run.

replicationRun_state :: Lens' ReplicationRun (Maybe ReplicationRunState)

The state of the replication run.

replicationRun_kmsKeyId :: Lens' ReplicationRun (Maybe Text)

The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is true but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

replicationRun_scheduledStartTime :: Lens' ReplicationRun (Maybe UTCTime)

The start time of the next replication run.

replicationRun_stageDetails :: Lens' ReplicationRun (Maybe ReplicationRunStageDetails)

Details about the current stage of the replication run.

replicationRun_description :: Lens' ReplicationRun (Maybe Text)

The description of the replication run.

replicationRun_type :: Lens' ReplicationRun (Maybe ReplicationRunType)

The type of replication run.