Network.AWS.SMS.Types.ServerGroupLaunchConfiguration

Description

Documentation

data ServerGroupLaunchConfiguration

Launch configuration for a server group.

See: newServerGroupLaunchConfiguration smart constructor.

Constructors

ServerGroupLaunchConfiguration'

Fields

serverGroupId :: Maybe Text

The ID of the server group with which the launch configuration is associated.

launchOrder :: Maybe Int

The launch order of servers in the server group.

serverLaunchConfigurations :: Maybe [ServerLaunchConfiguration]

The launch configuration for servers in the server group.

newServerGroupLaunchConfiguration :: ServerGroupLaunchConfiguration

Create a value of ServerGroupLaunchConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serverGroupId:ServerGroupLaunchConfiguration', serverGroupLaunchConfiguration_serverGroupId - The ID of the server group with which the launch configuration is associated.

$sel:launchOrder:ServerGroupLaunchConfiguration', serverGroupLaunchConfiguration_launchOrder - The launch order of servers in the server group.

$sel:serverLaunchConfigurations:ServerGroupLaunchConfiguration', serverGroupLaunchConfiguration_serverLaunchConfigurations - The launch configuration for servers in the server group.

serverGroupLaunchConfiguration_serverGroupId :: Lens' ServerGroupLaunchConfiguration (Maybe Text)

The ID of the server group with which the launch configuration is associated.

serverGroupLaunchConfiguration_launchOrder :: Lens' ServerGroupLaunchConfiguration (Maybe Int)

The launch order of servers in the server group.

serverGroupLaunchConfiguration_serverLaunchConfigurations :: Lens' ServerGroupLaunchConfiguration (Maybe [ServerLaunchConfiguration])

The launch configuration for servers in the server group.