Network.AWS.SMS.Types.ServerReplicationParameters

Description

Documentation

data ServerReplicationParameters

The replication parameters for replicating a server.

See: newServerReplicationParameters smart constructor.

Constructors

ServerReplicationParameters'

Fields

numberOfRecentAmisToKeep :: Maybe Int

The number of recent AMIs to keep when creating a replication job for this server.

encrypted :: Maybe Bool

Indicates whether the replication job produces encrypted AMIs.

seedTime :: Maybe POSIX

The seed time for creating a replication job for the server.

kmsKeyId :: Maybe Text

The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is enabled but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

frequency :: Maybe Int

The frequency of creating replication jobs for the server.

runOnce :: Maybe Bool

Indicates whether to run the replication job one time.

licenseType :: Maybe LicenseType

The license type for creating a replication job for the server.

newServerReplicationParameters :: ServerReplicationParameters

Create a value of ServerReplicationParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:numberOfRecentAmisToKeep:ServerReplicationParameters', serverReplicationParameters_numberOfRecentAmisToKeep - The number of recent AMIs to keep when creating a replication job for this server.

$sel:encrypted:ServerReplicationParameters', serverReplicationParameters_encrypted - Indicates whether the replication job produces encrypted AMIs.

$sel:seedTime:ServerReplicationParameters', serverReplicationParameters_seedTime - The seed time for creating a replication job for the server.

$sel:kmsKeyId:ServerReplicationParameters', serverReplicationParameters_kmsKeyId - The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is enabled but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

$sel:frequency:ServerReplicationParameters', serverReplicationParameters_frequency - The frequency of creating replication jobs for the server.

$sel:runOnce:ServerReplicationParameters', serverReplicationParameters_runOnce - Indicates whether to run the replication job one time.

$sel:licenseType:ServerReplicationParameters', serverReplicationParameters_licenseType - The license type for creating a replication job for the server.

serverReplicationParameters_numberOfRecentAmisToKeep :: Lens' ServerReplicationParameters (Maybe Int)

The number of recent AMIs to keep when creating a replication job for this server.

serverReplicationParameters_encrypted :: Lens' ServerReplicationParameters (Maybe Bool)

Indicates whether the replication job produces encrypted AMIs.

serverReplicationParameters_seedTime :: Lens' ServerReplicationParameters (Maybe UTCTime)

The seed time for creating a replication job for the server.

serverReplicationParameters_kmsKeyId :: Lens' ServerReplicationParameters (Maybe Text)

The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is enabled but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

serverReplicationParameters_frequency :: Lens' ServerReplicationParameters (Maybe Int)

The frequency of creating replication jobs for the server.

serverReplicationParameters_runOnce :: Lens' ServerReplicationParameters (Maybe Bool)

Indicates whether to run the replication job one time.

serverReplicationParameters_licenseType :: Lens' ServerReplicationParameters (Maybe LicenseType)

The license type for creating a replication job for the server.