Network.AWS.SMS.UpdateReplicationJob

Description

Updates the specified settings for the specified replication job.

Creating a Request

data UpdateReplicationJob

See: newUpdateReplicationJob smart constructor.

Constructors

UpdateReplicationJob'

Fields

frequency :: Maybe Int

The time between consecutive replication runs, in hours.

numberOfRecentAmisToKeep :: Maybe Int

The maximum number of SMS-created AMIs to retain. The oldest is deleted after the maximum number is reached and a new AMI is created.

licenseType :: Maybe LicenseType

The license type to be used for the AMI created by a successful replication run.

roleName :: Maybe Text

The name of the IAM role to be used by AWS SMS.

encrypted :: Maybe Bool

When true, the replication job produces encrypted AMIs. For more information, KmsKeyId.

nextReplicationRunStartTime :: Maybe POSIX

The start time of the next replication run.

kmsKeyId :: Maybe Text

The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is enabled but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

description :: Maybe Text

The description of the replication job.

replicationJobId :: Text

The ID of the replication job.

newUpdateReplicationJob

Arguments

:: Text
$sel:replicationJobId:UpdateReplicationJob'

-> UpdateReplicationJob

Create a value of UpdateReplicationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:frequency:UpdateReplicationJob', updateReplicationJob_frequency - The time between consecutive replication runs, in hours.

$sel:numberOfRecentAmisToKeep:UpdateReplicationJob', updateReplicationJob_numberOfRecentAmisToKeep - The maximum number of SMS-created AMIs to retain. The oldest is deleted after the maximum number is reached and a new AMI is created.

$sel:licenseType:UpdateReplicationJob', updateReplicationJob_licenseType - The license type to be used for the AMI created by a successful replication run.

$sel:roleName:UpdateReplicationJob', updateReplicationJob_roleName - The name of the IAM role to be used by AWS SMS.

$sel:encrypted:UpdateReplicationJob', updateReplicationJob_encrypted - When true, the replication job produces encrypted AMIs. For more information, KmsKeyId.

$sel:nextReplicationRunStartTime:UpdateReplicationJob', updateReplicationJob_nextReplicationRunStartTime - The start time of the next replication run.

$sel:kmsKeyId:UpdateReplicationJob', updateReplicationJob_kmsKeyId - The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is enabled but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

$sel:description:UpdateReplicationJob', updateReplicationJob_description - The description of the replication job.

$sel:replicationJobId:UpdateReplicationJob', updateReplicationJob_replicationJobId - The ID of the replication job.

Request Lenses

updateReplicationJob_frequency :: Lens' UpdateReplicationJob (Maybe Int)

The time between consecutive replication runs, in hours.

updateReplicationJob_numberOfRecentAmisToKeep :: Lens' UpdateReplicationJob (Maybe Int)

The maximum number of SMS-created AMIs to retain. The oldest is deleted after the maximum number is reached and a new AMI is created.

updateReplicationJob_licenseType :: Lens' UpdateReplicationJob (Maybe LicenseType)

The license type to be used for the AMI created by a successful replication run.

updateReplicationJob_roleName :: Lens' UpdateReplicationJob (Maybe Text)

The name of the IAM role to be used by AWS SMS.

updateReplicationJob_encrypted :: Lens' UpdateReplicationJob (Maybe Bool)

When true, the replication job produces encrypted AMIs. For more information, KmsKeyId.

updateReplicationJob_nextReplicationRunStartTime :: Lens' UpdateReplicationJob (Maybe UTCTime)

The start time of the next replication run.

updateReplicationJob_kmsKeyId :: Lens' UpdateReplicationJob (Maybe Text)

The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following:

KMS key ID

KMS key alias

ARN referring to the KMS key ID

ARN referring to the KMS key alias

If encrypted is enabled but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

updateReplicationJob_description :: Lens' UpdateReplicationJob (Maybe Text)

The description of the replication job.

updateReplicationJob_replicationJobId :: Lens' UpdateReplicationJob Text

The ID of the replication job.

Destructuring the Response

data UpdateReplicationJobResponse

See: newUpdateReplicationJobResponse smart constructor.

Constructors

UpdateReplicationJobResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateReplicationJobResponse

Arguments

:: Int
$sel:httpStatus:UpdateReplicationJobResponse'

-> UpdateReplicationJobResponse

Create a value of UpdateReplicationJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateReplicationJobResponse', updateReplicationJobResponse_httpStatus - The response's http status code.

Response Lenses

updateReplicationJobResponse_httpStatus :: Lens' UpdateReplicationJobResponse Int

The response's http status code.