Network.AWS.SNS.ConfirmSubscription

Description

Verifies an endpoint owner's intent to receive messages by validating the token sent to the endpoint by an earlier Subscribe action. If the token is valid, the action creates a new subscription and returns its Amazon Resource Name (ARN). This call requires an AWS signature only when the AuthenticateOnUnsubscribe flag is set to "true".

Creating a Request

data ConfirmSubscription

Input for ConfirmSubscription action.

See: newConfirmSubscription smart constructor.

Constructors

ConfirmSubscription'

Fields

authenticateOnUnsubscribe :: Maybe Text

Disallows unauthenticated unsubscribes of the subscription. If the value of this parameter is true and the request has an Amazon Web Services signature, then only the topic owner and the subscription owner can unsubscribe the endpoint. The unsubscribe action requires Amazon Web Services authentication.

topicArn :: Text

The ARN of the topic for which you wish to confirm a subscription.

token :: Text

Short-lived token sent to an endpoint during the Subscribe action.

newConfirmSubscription

Arguments

:: Text
$sel:topicArn:ConfirmSubscription'

-> Text
$sel:token:ConfirmSubscription'

-> ConfirmSubscription

Create a value of ConfirmSubscription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authenticateOnUnsubscribe:ConfirmSubscription', confirmSubscription_authenticateOnUnsubscribe - Disallows unauthenticated unsubscribes of the subscription. If the value of this parameter is true and the request has an Amazon Web Services signature, then only the topic owner and the subscription owner can unsubscribe the endpoint. The unsubscribe action requires Amazon Web Services authentication.

$sel:topicArn:ConfirmSubscription', confirmSubscription_topicArn - The ARN of the topic for which you wish to confirm a subscription.

$sel:token:ConfirmSubscription', confirmSubscription_token - Short-lived token sent to an endpoint during the Subscribe action.

Request Lenses

confirmSubscription_authenticateOnUnsubscribe :: Lens' ConfirmSubscription (Maybe Text)

Disallows unauthenticated unsubscribes of the subscription. If the value of this parameter is true and the request has an Amazon Web Services signature, then only the topic owner and the subscription owner can unsubscribe the endpoint. The unsubscribe action requires Amazon Web Services authentication.

confirmSubscription_topicArn :: Lens' ConfirmSubscription Text

The ARN of the topic for which you wish to confirm a subscription.

confirmSubscription_token :: Lens' ConfirmSubscription Text

Short-lived token sent to an endpoint during the Subscribe action.

Destructuring the Response

data ConfirmSubscriptionResponse

Response for ConfirmSubscriptions action.

See: newConfirmSubscriptionResponse smart constructor.

Constructors

ConfirmSubscriptionResponse'

Fields

subscriptionArn :: Maybe Text

The ARN of the created subscription.

httpStatus :: Int

The response's http status code.

newConfirmSubscriptionResponse

Arguments

:: Int
$sel:httpStatus:ConfirmSubscriptionResponse'

-> ConfirmSubscriptionResponse

Create a value of ConfirmSubscriptionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscriptionArn:ConfirmSubscriptionResponse', confirmSubscriptionResponse_subscriptionArn - The ARN of the created subscription.

$sel:httpStatus:ConfirmSubscriptionResponse', confirmSubscriptionResponse_httpStatus - The response's http status code.

Response Lenses

confirmSubscriptionResponse_subscriptionArn :: Lens' ConfirmSubscriptionResponse (Maybe Text)

The ARN of the created subscription.

confirmSubscriptionResponse_httpStatus :: Lens' ConfirmSubscriptionResponse Int

The response's http status code.