Network.AWS.SNS.CreatePlatformEndpoint

Description

Creates an endpoint for a device and mobile app on one of the supported push notification services, such as GCM (Firebase Cloud Messaging) and APNS. CreatePlatformEndpoint requires the PlatformApplicationArn that is returned from CreatePlatformApplication. You can use the returned EndpointArn to send a message to a mobile app or by the Subscribe action for subscription to a topic. The CreatePlatformEndpoint action is idempotent, so if the requester already owns an endpoint with the same device token and attributes, that endpoint's ARN is returned without creating a new endpoint. For more information, see Using Amazon SNS Mobile Push Notifications.

When using CreatePlatformEndpoint with Baidu, two attributes must be provided: ChannelId and UserId. The token field must also contain the ChannelId. For more information, see Creating an Amazon SNS Endpoint for Baidu.

Creating a Request

data CreatePlatformEndpoint

Input for CreatePlatformEndpoint action.

See: newCreatePlatformEndpoint smart constructor.

Constructors

CreatePlatformEndpoint'

Fields

customUserData :: Maybe Text

Arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB.

attributes :: Maybe (HashMap Text Text)

For a list of attributes, see SetEndpointAttributes.

platformApplicationArn :: Text

PlatformApplicationArn returned from CreatePlatformApplication is used to create a an endpoint.

token :: Text

Unique identifier created by the notification service for an app on a device. The specific name for Token will vary, depending on which notification service is being used. For example, when using APNS as the notification service, you need the device token. Alternatively, when using GCM (Firebase Cloud Messaging) or ADM, the device token equivalent is called the registration ID.

newCreatePlatformEndpoint

Arguments

:: Text
$sel:platformApplicationArn:CreatePlatformEndpoint'

-> Text
$sel:token:CreatePlatformEndpoint'

-> CreatePlatformEndpoint

Create a value of CreatePlatformEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customUserData:CreatePlatformEndpoint', createPlatformEndpoint_customUserData - Arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB.

$sel:attributes:CreatePlatformEndpoint', createPlatformEndpoint_attributes - For a list of attributes, see SetEndpointAttributes.

$sel:platformApplicationArn:CreatePlatformEndpoint', createPlatformEndpoint_platformApplicationArn - PlatformApplicationArn returned from CreatePlatformApplication is used to create a an endpoint.

$sel:token:CreatePlatformEndpoint', createPlatformEndpoint_token - Unique identifier created by the notification service for an app on a device. The specific name for Token will vary, depending on which notification service is being used. For example, when using APNS as the notification service, you need the device token. Alternatively, when using GCM (Firebase Cloud Messaging) or ADM, the device token equivalent is called the registration ID.

Request Lenses

createPlatformEndpoint_customUserData :: Lens' CreatePlatformEndpoint (Maybe Text)

Arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB.

createPlatformEndpoint_attributes :: Lens' CreatePlatformEndpoint (Maybe (HashMap Text Text))

For a list of attributes, see SetEndpointAttributes.

createPlatformEndpoint_platformApplicationArn :: Lens' CreatePlatformEndpoint Text

PlatformApplicationArn returned from CreatePlatformApplication is used to create a an endpoint.

createPlatformEndpoint_token :: Lens' CreatePlatformEndpoint Text

Unique identifier created by the notification service for an app on a device. The specific name for Token will vary, depending on which notification service is being used. For example, when using APNS as the notification service, you need the device token. Alternatively, when using GCM (Firebase Cloud Messaging) or ADM, the device token equivalent is called the registration ID.

Destructuring the Response

data CreatePlatformEndpointResponse

Response from CreateEndpoint action.

See: newCreatePlatformEndpointResponse smart constructor.

Constructors

CreatePlatformEndpointResponse'

Fields

endpointArn :: Maybe Text

EndpointArn returned from CreateEndpoint action.

httpStatus :: Int

The response's http status code.

newCreatePlatformEndpointResponse

Arguments

:: Int
$sel:httpStatus:CreatePlatformEndpointResponse'

-> CreatePlatformEndpointResponse

Create a value of CreatePlatformEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpointArn:CreatePlatformEndpointResponse', createPlatformEndpointResponse_endpointArn - EndpointArn returned from CreateEndpoint action.

$sel:httpStatus:CreatePlatformEndpointResponse', createPlatformEndpointResponse_httpStatus - The response's http status code.

Response Lenses

createPlatformEndpointResponse_endpointArn :: Lens' CreatePlatformEndpointResponse (Maybe Text)

EndpointArn returned from CreateEndpoint action.

createPlatformEndpointResponse_httpStatus :: Lens' CreatePlatformEndpointResponse Int

The response's http status code.