Network.AWS.SNS.CreateSMSSandboxPhoneNumber

Description

Adds a destination phone number to an account in the SMS sandbox and sends a one-time password (OTP) to that phone number.

When you start using Amazon SNS to send SMS messages, your account is in the SMS sandbox. The SMS sandbox provides a safe environment for you to try Amazon SNS features without risking your reputation as an SMS sender. While your account is in the SMS sandbox, you can use all of the features of Amazon SNS. However, you can send SMS messages only to verified destination phone numbers. For more information, including how to move out of the sandbox to send messages without restrictions, see SMS sandbox in the Amazon SNS Developer Guide.

Creating a Request

data CreateSMSSandboxPhoneNumber

See: newCreateSMSSandboxPhoneNumber smart constructor.

Constructors

CreateSMSSandboxPhoneNumber'

Fields

languageCode :: Maybe LanguageCodeString

The language to use for sending the OTP. The default value is en-US.

phoneNumber :: Text

The destination phone number to verify. On verification, Amazon SNS adds this phone number to the list of verified phone numbers that you can send SMS messages to.

newCreateSMSSandboxPhoneNumber

Arguments

:: Text
$sel:phoneNumber:CreateSMSSandboxPhoneNumber'

-> CreateSMSSandboxPhoneNumber

Create a value of CreateSMSSandboxPhoneNumber with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:CreateSMSSandboxPhoneNumber', createSMSSandboxPhoneNumber_languageCode - The language to use for sending the OTP. The default value is en-US.

$sel:phoneNumber:CreateSMSSandboxPhoneNumber', createSMSSandboxPhoneNumber_phoneNumber - The destination phone number to verify. On verification, Amazon SNS adds this phone number to the list of verified phone numbers that you can send SMS messages to.

Request Lenses

createSMSSandboxPhoneNumber_languageCode :: Lens' CreateSMSSandboxPhoneNumber (Maybe LanguageCodeString)

The language to use for sending the OTP. The default value is en-US.

createSMSSandboxPhoneNumber_phoneNumber :: Lens' CreateSMSSandboxPhoneNumber Text

The destination phone number to verify. On verification, Amazon SNS adds this phone number to the list of verified phone numbers that you can send SMS messages to.

Destructuring the Response

data CreateSMSSandboxPhoneNumberResponse

See: newCreateSMSSandboxPhoneNumberResponse smart constructor.

Constructors

CreateSMSSandboxPhoneNumberResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateSMSSandboxPhoneNumberResponse

Arguments

:: Int
$sel:httpStatus:CreateSMSSandboxPhoneNumberResponse'

-> CreateSMSSandboxPhoneNumberResponse

Create a value of CreateSMSSandboxPhoneNumberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateSMSSandboxPhoneNumberResponse', createSMSSandboxPhoneNumberResponse_httpStatus - The response's http status code.

Response Lenses

createSMSSandboxPhoneNumberResponse_httpStatus :: Lens' CreateSMSSandboxPhoneNumberResponse Int

The response's http status code.