Network.AWS.SNS.CreateTopic

Description

Creates a topic to which notifications can be published. Users can create at most 100,000 standard topics (at most 1,000 FIFO topics). For more information, see Creating an Amazon SNS topic in the Amazon SNS Developer Guide. This action is idempotent, so if the requester already owns a topic with the specified name, that topic's ARN is returned without creating a new topic.

Creating a Request

data CreateTopic

Input for CreateTopic action.

See: newCreateTopic smart constructor.

Constructors

CreateTopic'

Fields

attributes :: Maybe (HashMap Text Text)

A map of attributes with their corresponding values.

The following lists the names, descriptions, and values of the special request parameters that the CreateTopic action uses:

DeliveryPolicy – The policy that defines how Amazon SNS retries failed deliveries to HTTP/S endpoints.

DisplayName – The display name to use for a topic with SMS subscriptions.

FifoTopic – Set to true to create a FIFO topic.

Policy – The policy that defines who can access your topic. By default, only the topic owner can publish or subscribe to the topic.

The following attribute applies only to server-side encryption:

KmsMasterKeyId – The ID of an Amazon Web Services managed customer master key (CMK) for Amazon SNS or a custom CMK. For more information, see Key Terms. For more examples, see KeyId in the Key Management Service API Reference.

The following attributes apply only to FIFO topics:

FifoTopic – When this is set to true, a FIFO topic is created.

ContentBasedDeduplication – Enables content-based deduplication for FIFO topics.

By default, ContentBasedDeduplication is set to false. If you create a FIFO topic and this attribute is false, you must specify a value for the MessageDeduplicationId parameter for the Publish action.

When you set ContentBasedDeduplication to true, Amazon SNS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message).

(Optional) To override the generated value, you can specify a value for the MessageDeduplicationId parameter for the Publish action.

tags :: Maybe [Tag]

The list of tags to add to a new topic.

To be able to tag a topic on creation, you must have the sns:CreateTopic and sns:TagResource permissions.

name :: Text

The name of the topic you want to create.

Constraints: Topic names must be made up of only uppercase and lowercase ASCII letters, numbers, underscores, and hyphens, and must be between 1 and 256 characters long.

For a FIFO (first-in-first-out) topic, the name must end with the .fifo suffix.

newCreateTopic

Arguments

:: Text
$sel:name:CreateTopic'

-> CreateTopic

Create a value of CreateTopic with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:CreateTopic', createTopic_attributes - A map of attributes with their corresponding values.

The following lists the names, descriptions, and values of the special request parameters that the CreateTopic action uses:

DeliveryPolicy – The policy that defines how Amazon SNS retries failed deliveries to HTTP/S endpoints.

DisplayName – The display name to use for a topic with SMS subscriptions.

FifoTopic – Set to true to create a FIFO topic.

Policy – The policy that defines who can access your topic. By default, only the topic owner can publish or subscribe to the topic.

The following attribute applies only to server-side encryption:

KmsMasterKeyId – The ID of an Amazon Web Services managed customer master key (CMK) for Amazon SNS or a custom CMK. For more information, see Key Terms. For more examples, see KeyId in the Key Management Service API Reference.

The following attributes apply only to FIFO topics:

FifoTopic – When this is set to true, a FIFO topic is created.

ContentBasedDeduplication – Enables content-based deduplication for FIFO topics.

By default, ContentBasedDeduplication is set to false. If you create a FIFO topic and this attribute is false, you must specify a value for the MessageDeduplicationId parameter for the Publish action.

When you set ContentBasedDeduplication to true, Amazon SNS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message).

(Optional) To override the generated value, you can specify a value for the MessageDeduplicationId parameter for the Publish action.

$sel:tags:CreateTopic', createTopic_tags - The list of tags to add to a new topic.

To be able to tag a topic on creation, you must have the sns:CreateTopic and sns:TagResource permissions.

$sel:name:CreateTopic', createTopic_name - The name of the topic you want to create.

Constraints: Topic names must be made up of only uppercase and lowercase ASCII letters, numbers, underscores, and hyphens, and must be between 1 and 256 characters long.

For a FIFO (first-in-first-out) topic, the name must end with the .fifo suffix.

Request Lenses

createTopic_attributes :: Lens' CreateTopic (Maybe (HashMap Text Text))

A map of attributes with their corresponding values.

The following lists the names, descriptions, and values of the special request parameters that the CreateTopic action uses:

DeliveryPolicy – The policy that defines how Amazon SNS retries failed deliveries to HTTP/S endpoints.

DisplayName – The display name to use for a topic with SMS subscriptions.

FifoTopic – Set to true to create a FIFO topic.

Policy – The policy that defines who can access your topic. By default, only the topic owner can publish or subscribe to the topic.

The following attribute applies only to server-side encryption:

KmsMasterKeyId – The ID of an Amazon Web Services managed customer master key (CMK) for Amazon SNS or a custom CMK. For more information, see Key Terms. For more examples, see KeyId in the Key Management Service API Reference.

The following attributes apply only to FIFO topics:

FifoTopic – When this is set to true, a FIFO topic is created.

ContentBasedDeduplication – Enables content-based deduplication for FIFO topics.

By default, ContentBasedDeduplication is set to false. If you create a FIFO topic and this attribute is false, you must specify a value for the MessageDeduplicationId parameter for the Publish action.

When you set ContentBasedDeduplication to true, Amazon SNS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message).

(Optional) To override the generated value, you can specify a value for the MessageDeduplicationId parameter for the Publish action.

createTopic_tags :: Lens' CreateTopic (Maybe [Tag])

The list of tags to add to a new topic.

To be able to tag a topic on creation, you must have the sns:CreateTopic and sns:TagResource permissions.

createTopic_name :: Lens' CreateTopic Text

The name of the topic you want to create.

Constraints: Topic names must be made up of only uppercase and lowercase ASCII letters, numbers, underscores, and hyphens, and must be between 1 and 256 characters long.

For a FIFO (first-in-first-out) topic, the name must end with the .fifo suffix.

Destructuring the Response

data CreateTopicResponse

Response from CreateTopic action.

See: newCreateTopicResponse smart constructor.

Constructors

CreateTopicResponse'

Fields

topicArn :: Maybe Text

The Amazon Resource Name (ARN) assigned to the created topic.

httpStatus :: Int

The response's http status code.

newCreateTopicResponse

Arguments

:: Int
$sel:httpStatus:CreateTopicResponse'

-> CreateTopicResponse

Create a value of CreateTopicResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:topicArn:CreateTopicResponse', createTopicResponse_topicArn - The Amazon Resource Name (ARN) assigned to the created topic.

$sel:httpStatus:CreateTopicResponse', createTopicResponse_httpStatus - The response's http status code.

Response Lenses

createTopicResponse_topicArn :: Lens' CreateTopicResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the created topic.

createTopicResponse_httpStatus :: Lens' CreateTopicResponse Int

The response's http status code.