Network.AWS.SNS.GetEndpointAttributes

Description

Retrieves the endpoint attributes for a device on one of the supported push notification services, such as GCM (Firebase Cloud Messaging) and APNS. For more information, see Using Amazon SNS Mobile Push Notifications.

Creating a Request

data GetEndpointAttributes

Input for GetEndpointAttributes action.

See: newGetEndpointAttributes smart constructor.

Constructors

GetEndpointAttributes'

Fields

endpointArn :: Text

EndpointArn for GetEndpointAttributes input.

newGetEndpointAttributes

Arguments

:: Text
$sel:endpointArn:GetEndpointAttributes'

-> GetEndpointAttributes

Create a value of GetEndpointAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpointArn:GetEndpointAttributes', getEndpointAttributes_endpointArn - EndpointArn for GetEndpointAttributes input.

Request Lenses

getEndpointAttributes_endpointArn :: Lens' GetEndpointAttributes Text

EndpointArn for GetEndpointAttributes input.

Destructuring the Response

data GetEndpointAttributesResponse

Response from GetEndpointAttributes of the EndpointArn.

See: newGetEndpointAttributesResponse smart constructor.

Constructors

GetEndpointAttributesResponse'

Fields

attributes :: Maybe (HashMap Text Text)

Attributes include the following:

CustomUserData – arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB.

Enabled – flag that enables/disables delivery to the endpoint. Amazon SNS will set this to false when a notification service indicates to Amazon SNS that the endpoint is invalid. Users can set it back to true, typically after updating Token.

Token – device token, also referred to as a registration id, for an app and mobile device. This is returned from the notification service when an app and mobile device are registered with the notification service.

The device token for the iOS platform is returned in lowercase.

httpStatus :: Int

The response's http status code.

newGetEndpointAttributesResponse

Arguments

:: Int
$sel:httpStatus:GetEndpointAttributesResponse'

-> GetEndpointAttributesResponse

Create a value of GetEndpointAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:GetEndpointAttributesResponse', getEndpointAttributesResponse_attributes - Attributes include the following:

CustomUserData – arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB.

Enabled – flag that enables/disables delivery to the endpoint. Amazon SNS will set this to false when a notification service indicates to Amazon SNS that the endpoint is invalid. Users can set it back to true, typically after updating Token.

Token – device token, also referred to as a registration id, for an app and mobile device. This is returned from the notification service when an app and mobile device are registered with the notification service.

The device token for the iOS platform is returned in lowercase.

$sel:httpStatus:GetEndpointAttributesResponse', getEndpointAttributesResponse_httpStatus - The response's http status code.

Response Lenses

getEndpointAttributesResponse_attributes :: Lens' GetEndpointAttributesResponse (Maybe (HashMap Text Text))

Attributes include the following:

CustomUserData – arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB.

Enabled – flag that enables/disables delivery to the endpoint. Amazon SNS will set this to false when a notification service indicates to Amazon SNS that the endpoint is invalid. Users can set it back to true, typically after updating Token.

Token – device token, also referred to as a registration id, for an app and mobile device. This is returned from the notification service when an app and mobile device are registered with the notification service.

The device token for the iOS platform is returned in lowercase.

getEndpointAttributesResponse_httpStatus :: Lens' GetEndpointAttributesResponse Int

The response's http status code.