Network.AWS.SNS.GetPlatformApplicationAttributes

Description

Retrieves the attributes of the platform application object for the supported push notification services, such as APNS and GCM (Firebase Cloud Messaging). For more information, see Using Amazon SNS Mobile Push Notifications.

Creating a Request

data GetPlatformApplicationAttributes

Input for GetPlatformApplicationAttributes action.

See: newGetPlatformApplicationAttributes smart constructor.

Constructors

GetPlatformApplicationAttributes'

Fields

platformApplicationArn :: Text

PlatformApplicationArn for GetPlatformApplicationAttributesInput.

newGetPlatformApplicationAttributes

Arguments

:: Text
$sel:platformApplicationArn:GetPlatformApplicationAttributes'

-> GetPlatformApplicationAttributes

Create a value of GetPlatformApplicationAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platformApplicationArn:GetPlatformApplicationAttributes', getPlatformApplicationAttributes_platformApplicationArn - PlatformApplicationArn for GetPlatformApplicationAttributesInput.

Request Lenses

getPlatformApplicationAttributes_platformApplicationArn :: Lens' GetPlatformApplicationAttributes Text

PlatformApplicationArn for GetPlatformApplicationAttributesInput.

Destructuring the Response

data GetPlatformApplicationAttributesResponse

Response for GetPlatformApplicationAttributes action.

See: newGetPlatformApplicationAttributesResponse smart constructor.

Constructors

GetPlatformApplicationAttributesResponse'

Fields

attributes :: Maybe (HashMap Text Text)

Attributes include the following:

EventEndpointCreated – Topic ARN to which EndpointCreated event notifications should be sent.

EventEndpointDeleted – Topic ARN to which EndpointDeleted event notifications should be sent.

EventEndpointUpdated – Topic ARN to which EndpointUpdate event notifications should be sent.

EventDeliveryFailure – Topic ARN to which DeliveryFailure event notifications should be sent upon Direct Publish delivery failure (permanent) to one of the application's endpoints.

httpStatus :: Int

The response's http status code.

newGetPlatformApplicationAttributesResponse

Arguments

:: Int
$sel:httpStatus:GetPlatformApplicationAttributesResponse'

-> GetPlatformApplicationAttributesResponse

Create a value of GetPlatformApplicationAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:GetPlatformApplicationAttributesResponse', getPlatformApplicationAttributesResponse_attributes - Attributes include the following:

EventEndpointCreated – Topic ARN to which EndpointCreated event notifications should be sent.

EventEndpointDeleted – Topic ARN to which EndpointDeleted event notifications should be sent.

EventEndpointUpdated – Topic ARN to which EndpointUpdate event notifications should be sent.

EventDeliveryFailure – Topic ARN to which DeliveryFailure event notifications should be sent upon Direct Publish delivery failure (permanent) to one of the application's endpoints.

$sel:httpStatus:GetPlatformApplicationAttributesResponse', getPlatformApplicationAttributesResponse_httpStatus - The response's http status code.

Response Lenses

getPlatformApplicationAttributesResponse_attributes :: Lens' GetPlatformApplicationAttributesResponse (Maybe (HashMap Text Text))

Attributes include the following:

EventEndpointCreated – Topic ARN to which EndpointCreated event notifications should be sent.

EventEndpointDeleted – Topic ARN to which EndpointDeleted event notifications should be sent.

EventEndpointUpdated – Topic ARN to which EndpointUpdate event notifications should be sent.

EventDeliveryFailure – Topic ARN to which DeliveryFailure event notifications should be sent upon Direct Publish delivery failure (permanent) to one of the application's endpoints.

getPlatformApplicationAttributesResponse_httpStatus :: Lens' GetPlatformApplicationAttributesResponse Int

The response's http status code.