Network.AWS.SNS.GetSMSSandboxAccountStatus

Description

Retrieves the SMS sandbox status for the calling account in the target Region.

When you start using Amazon SNS to send SMS messages, your account is in the SMS sandbox. The SMS sandbox provides a safe environment for you to try Amazon SNS features without risking your reputation as an SMS sender. While your account is in the SMS sandbox, you can use all of the features of Amazon SNS. However, you can send SMS messages only to verified destination phone numbers. For more information, including how to move out of the sandbox to send messages without restrictions, see SMS sandbox in the Amazon SNS Developer Guide.

Creating a Request

data GetSMSSandboxAccountStatus

See: newGetSMSSandboxAccountStatus smart constructor.

Constructors

GetSMSSandboxAccountStatus'

newGetSMSSandboxAccountStatus :: GetSMSSandboxAccountStatus

Create a value of GetSMSSandboxAccountStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetSMSSandboxAccountStatusResponse

See: newGetSMSSandboxAccountStatusResponse smart constructor.

Constructors

GetSMSSandboxAccountStatusResponse'

Fields

httpStatus :: Int

The response's http status code.

isInSandbox :: Bool

Indicates whether the calling account is in the SMS sandbox.

newGetSMSSandboxAccountStatusResponse

Arguments

:: Int
$sel:httpStatus:GetSMSSandboxAccountStatusResponse'

-> Bool
$sel:isInSandbox:GetSMSSandboxAccountStatusResponse'

-> GetSMSSandboxAccountStatusResponse

Create a value of GetSMSSandboxAccountStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetSMSSandboxAccountStatusResponse', getSMSSandboxAccountStatusResponse_httpStatus - The response's http status code.

$sel:isInSandbox:GetSMSSandboxAccountStatusResponse', getSMSSandboxAccountStatusResponse_isInSandbox - Indicates whether the calling account is in the SMS sandbox.

Response Lenses

getSMSSandboxAccountStatusResponse_httpStatus :: Lens' GetSMSSandboxAccountStatusResponse Int

The response's http status code.

getSMSSandboxAccountStatusResponse_isInSandbox :: Lens' GetSMSSandboxAccountStatusResponse Bool

Indicates whether the calling account is in the SMS sandbox.