Network.AWS.SNS.GetTopicAttributes

Description

Returns all of the properties of a topic. Topic properties returned might differ based on the authorization of the user.

Creating a Request

data GetTopicAttributes

Input for GetTopicAttributes action.

See: newGetTopicAttributes smart constructor.

Constructors

GetTopicAttributes'

Fields

topicArn :: Text

The ARN of the topic whose properties you want to get.

newGetTopicAttributes

Arguments

:: Text
$sel:topicArn:GetTopicAttributes'

-> GetTopicAttributes

Create a value of GetTopicAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:topicArn:GetTopicAttributes', getTopicAttributes_topicArn - The ARN of the topic whose properties you want to get.

Request Lenses

getTopicAttributes_topicArn :: Lens' GetTopicAttributes Text

The ARN of the topic whose properties you want to get.

Destructuring the Response

data GetTopicAttributesResponse

Response for GetTopicAttributes action.

See: newGetTopicAttributesResponse smart constructor.

Constructors

GetTopicAttributesResponse'

Fields

attributes :: Maybe (HashMap Text Text)

A map of the topic's attributes. Attributes in this map include the following:

DeliveryPolicy – The JSON serialization of the topic's delivery policy.

DisplayName – The human-readable name used in the From field for notifications to email and email-json endpoints.

Owner – The account ID of the topic's owner.

Policy – The JSON serialization of the topic's access control policy.

SubscriptionsConfirmed – The number of confirmed subscriptions for the topic.

SubscriptionsDeleted – The number of deleted subscriptions for the topic.

SubscriptionsPending – The number of subscriptions pending confirmation for the topic.

TopicArn – The topic's ARN.

EffectiveDeliveryPolicy – The JSON serialization of the effective delivery policy, taking system defaults into account.

The following attribute applies only to server-side-encryption:

KmsMasterKeyId - The ID of an Amazon Web Services managed customer master key (CMK) for Amazon SNS or a custom CMK. For more information, see Key Terms. For more examples, see KeyId in the Key Management Service API Reference.

The following attributes apply only to FIFO topics:

FifoTopic – When this is set to true, a FIFO topic is created.

ContentBasedDeduplication – Enables content-based deduplication for FIFO topics.

By default, ContentBasedDeduplication is set to false. If you create a FIFO topic and this attribute is false, you must specify a value for the MessageDeduplicationId parameter for the Publish action.

When you set ContentBasedDeduplication to true, Amazon SNS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message).

(Optional) To override the generated value, you can specify a value for the MessageDeduplicationId parameter for the Publish action.

httpStatus :: Int

The response's http status code.

newGetTopicAttributesResponse

Arguments

:: Int
$sel:httpStatus:GetTopicAttributesResponse'

-> GetTopicAttributesResponse

Create a value of GetTopicAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:GetTopicAttributesResponse', getTopicAttributesResponse_attributes - A map of the topic's attributes. Attributes in this map include the following:

DeliveryPolicy – The JSON serialization of the topic's delivery policy.

DisplayName – The human-readable name used in the From field for notifications to email and email-json endpoints.

Owner – The account ID of the topic's owner.

Policy – The JSON serialization of the topic's access control policy.

SubscriptionsConfirmed – The number of confirmed subscriptions for the topic.

SubscriptionsDeleted – The number of deleted subscriptions for the topic.

SubscriptionsPending – The number of subscriptions pending confirmation for the topic.

TopicArn – The topic's ARN.

EffectiveDeliveryPolicy – The JSON serialization of the effective delivery policy, taking system defaults into account.

The following attribute applies only to server-side-encryption:

KmsMasterKeyId - The ID of an Amazon Web Services managed customer master key (CMK) for Amazon SNS or a custom CMK. For more information, see Key Terms. For more examples, see KeyId in the Key Management Service API Reference.

The following attributes apply only to FIFO topics:

FifoTopic – When this is set to true, a FIFO topic is created.

ContentBasedDeduplication – Enables content-based deduplication for FIFO topics.

By default, ContentBasedDeduplication is set to false. If you create a FIFO topic and this attribute is false, you must specify a value for the MessageDeduplicationId parameter for the Publish action.

When you set ContentBasedDeduplication to true, Amazon SNS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message).

(Optional) To override the generated value, you can specify a value for the MessageDeduplicationId parameter for the Publish action.

$sel:httpStatus:GetTopicAttributesResponse', getTopicAttributesResponse_httpStatus - The response's http status code.

Response Lenses

getTopicAttributesResponse_attributes :: Lens' GetTopicAttributesResponse (Maybe (HashMap Text Text))

A map of the topic's attributes. Attributes in this map include the following:

DeliveryPolicy – The JSON serialization of the topic's delivery policy.

DisplayName – The human-readable name used in the From field for notifications to email and email-json endpoints.

Owner – The account ID of the topic's owner.

Policy – The JSON serialization of the topic's access control policy.

SubscriptionsConfirmed – The number of confirmed subscriptions for the topic.

SubscriptionsDeleted – The number of deleted subscriptions for the topic.

SubscriptionsPending – The number of subscriptions pending confirmation for the topic.

TopicArn – The topic's ARN.

EffectiveDeliveryPolicy – The JSON serialization of the effective delivery policy, taking system defaults into account.

The following attribute applies only to server-side-encryption:

KmsMasterKeyId - The ID of an Amazon Web Services managed customer master key (CMK) for Amazon SNS or a custom CMK. For more information, see Key Terms. For more examples, see KeyId in the Key Management Service API Reference.

The following attributes apply only to FIFO topics:

FifoTopic – When this is set to true, a FIFO topic is created.

ContentBasedDeduplication – Enables content-based deduplication for FIFO topics.

By default, ContentBasedDeduplication is set to false. If you create a FIFO topic and this attribute is false, you must specify a value for the MessageDeduplicationId parameter for the Publish action.

When you set ContentBasedDeduplication to true, Amazon SNS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message).

(Optional) To override the generated value, you can specify a value for the MessageDeduplicationId parameter for the Publish action.

getTopicAttributesResponse_httpStatus :: Lens' GetTopicAttributesResponse Int

The response's http status code.