Network.AWS.SNS.ListPlatformApplications

Description

Lists the platform application objects for the supported push notification services, such as APNS and GCM (Firebase Cloud Messaging). The results for ListPlatformApplications are paginated and return a limited list of applications, up to 100. If additional records are available after the first page results, then a NextToken string will be returned. To receive the next page, you call ListPlatformApplications using the NextToken string received from the previous call. When there are no more records to return, NextToken will be null. For more information, see Using Amazon SNS Mobile Push Notifications.

This action is throttled at 15 transactions per second (TPS).

This operation returns paginated results.

Creating a Request

data ListPlatformApplications

Input for ListPlatformApplications action.

See: newListPlatformApplications smart constructor.

Constructors

ListPlatformApplications'

Fields

nextToken :: Maybe Text

NextToken string is used when calling ListPlatformApplications action to retrieve additional records that are available after the first page results.

newListPlatformApplications :: ListPlatformApplications

Create a value of ListPlatformApplications with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPlatformApplications', listPlatformApplications_nextToken - NextToken string is used when calling ListPlatformApplications action to retrieve additional records that are available after the first page results.

Request Lenses

listPlatformApplications_nextToken :: Lens' ListPlatformApplications (Maybe Text)

NextToken string is used when calling ListPlatformApplications action to retrieve additional records that are available after the first page results.

Destructuring the Response

data ListPlatformApplicationsResponse

Response for ListPlatformApplications action.

See: newListPlatformApplicationsResponse smart constructor.

Constructors

ListPlatformApplicationsResponse'

Fields

nextToken :: Maybe Text

NextToken string is returned when calling ListPlatformApplications action if additional records are available after the first page results.

platformApplications :: Maybe [PlatformApplication]

Platform applications returned when calling ListPlatformApplications action.

httpStatus :: Int

The response's http status code.

newListPlatformApplicationsResponse

Arguments

:: Int
$sel:httpStatus:ListPlatformApplicationsResponse'

-> ListPlatformApplicationsResponse

Create a value of ListPlatformApplicationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPlatformApplications', listPlatformApplicationsResponse_nextToken - NextToken string is returned when calling ListPlatformApplications action if additional records are available after the first page results.

$sel:platformApplications:ListPlatformApplicationsResponse', listPlatformApplicationsResponse_platformApplications - Platform applications returned when calling ListPlatformApplications action.

$sel:httpStatus:ListPlatformApplicationsResponse', listPlatformApplicationsResponse_httpStatus - The response's http status code.

Response Lenses

listPlatformApplicationsResponse_nextToken :: Lens' ListPlatformApplicationsResponse (Maybe Text)

NextToken string is returned when calling ListPlatformApplications action if additional records are available after the first page results.

listPlatformApplicationsResponse_platformApplications :: Lens' ListPlatformApplicationsResponse (Maybe [PlatformApplication])

Platform applications returned when calling ListPlatformApplications action.

listPlatformApplicationsResponse_httpStatus :: Lens' ListPlatformApplicationsResponse Int

The response's http status code.