Network.AWS.SNS.ListSMSSandboxPhoneNumbers

Description

Lists the calling account's current verified and pending destination phone numbers in the SMS sandbox.

When you start using Amazon SNS to send SMS messages, your account is in the SMS sandbox. The SMS sandbox provides a safe environment for you to try Amazon SNS features without risking your reputation as an SMS sender. While your account is in the SMS sandbox, you can use all of the features of Amazon SNS. However, you can send SMS messages only to verified destination phone numbers. For more information, including how to move out of the sandbox to send messages without restrictions, see SMS sandbox in the Amazon SNS Developer Guide.

This operation returns paginated results.

Creating a Request

data ListSMSSandboxPhoneNumbers

See: newListSMSSandboxPhoneNumbers smart constructor.

Constructors

ListSMSSandboxPhoneNumbers'

Fields

nextToken :: Maybe Text

Token that the previous ListSMSSandboxPhoneNumbersInput request returns.

maxResults :: Maybe Natural

The maximum number of phone numbers to return.

newListSMSSandboxPhoneNumbers :: ListSMSSandboxPhoneNumbers

Create a value of ListSMSSandboxPhoneNumbers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSMSSandboxPhoneNumbers', listSMSSandboxPhoneNumbers_nextToken - Token that the previous ListSMSSandboxPhoneNumbersInput request returns.

$sel:maxResults:ListSMSSandboxPhoneNumbers', listSMSSandboxPhoneNumbers_maxResults - The maximum number of phone numbers to return.

Request Lenses

listSMSSandboxPhoneNumbers_nextToken :: Lens' ListSMSSandboxPhoneNumbers (Maybe Text)

Token that the previous ListSMSSandboxPhoneNumbersInput request returns.

listSMSSandboxPhoneNumbers_maxResults :: Lens' ListSMSSandboxPhoneNumbers (Maybe Natural)

The maximum number of phone numbers to return.

Destructuring the Response

data ListSMSSandboxPhoneNumbersResponse

See: newListSMSSandboxPhoneNumbersResponse smart constructor.

Constructors

ListSMSSandboxPhoneNumbersResponse'

Fields

nextToken :: Maybe Text

A NextToken string is returned when you call the ListSMSSandboxPhoneNumbersInput operation if additional pages of records are available.

httpStatus :: Int

The response's http status code.

phoneNumbers :: [SMSSandboxPhoneNumber]

A list of the calling account's pending and verified phone numbers.

newListSMSSandboxPhoneNumbersResponse

Arguments

:: Int
$sel:httpStatus:ListSMSSandboxPhoneNumbersResponse'

-> ListSMSSandboxPhoneNumbersResponse

Create a value of ListSMSSandboxPhoneNumbersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSMSSandboxPhoneNumbers', listSMSSandboxPhoneNumbersResponse_nextToken - A NextToken string is returned when you call the ListSMSSandboxPhoneNumbersInput operation if additional pages of records are available.

$sel:httpStatus:ListSMSSandboxPhoneNumbersResponse', listSMSSandboxPhoneNumbersResponse_httpStatus - The response's http status code.

$sel:phoneNumbers:ListSMSSandboxPhoneNumbersResponse', listSMSSandboxPhoneNumbersResponse_phoneNumbers - A list of the calling account's pending and verified phone numbers.

Response Lenses

listSMSSandboxPhoneNumbersResponse_nextToken :: Lens' ListSMSSandboxPhoneNumbersResponse (Maybe Text)

A NextToken string is returned when you call the ListSMSSandboxPhoneNumbersInput operation if additional pages of records are available.

listSMSSandboxPhoneNumbersResponse_httpStatus :: Lens' ListSMSSandboxPhoneNumbersResponse Int

The response's http status code.

listSMSSandboxPhoneNumbersResponse_phoneNumbers :: Lens' ListSMSSandboxPhoneNumbersResponse [SMSSandboxPhoneNumber]

A list of the calling account's pending and verified phone numbers.