Network.AWS.SNS.ListSubscriptionsByTopic

Description

Returns a list of the subscriptions to a specific topic. Each call returns a limited list of subscriptions, up to 100. If there are more subscriptions, a NextToken is also returned. Use the NextToken parameter in a new ListSubscriptionsByTopic call to get further results.

This action is throttled at 30 transactions per second (TPS).

This operation returns paginated results.

Creating a Request

data ListSubscriptionsByTopic

Input for ListSubscriptionsByTopic action.

See: newListSubscriptionsByTopic smart constructor.

Constructors

ListSubscriptionsByTopic'

Fields

nextToken :: Maybe Text

Token returned by the previous ListSubscriptionsByTopic request.

topicArn :: Text

The ARN of the topic for which you wish to find subscriptions.

newListSubscriptionsByTopic

Arguments

:: Text
$sel:topicArn:ListSubscriptionsByTopic'

-> ListSubscriptionsByTopic

Create a value of ListSubscriptionsByTopic with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSubscriptionsByTopic', listSubscriptionsByTopic_nextToken - Token returned by the previous ListSubscriptionsByTopic request.

$sel:topicArn:ListSubscriptionsByTopic', listSubscriptionsByTopic_topicArn - The ARN of the topic for which you wish to find subscriptions.

Request Lenses

listSubscriptionsByTopic_nextToken :: Lens' ListSubscriptionsByTopic (Maybe Text)

Token returned by the previous ListSubscriptionsByTopic request.

listSubscriptionsByTopic_topicArn :: Lens' ListSubscriptionsByTopic Text

The ARN of the topic for which you wish to find subscriptions.

Destructuring the Response

data ListSubscriptionsByTopicResponse

Response for ListSubscriptionsByTopic action.

See: newListSubscriptionsByTopicResponse smart constructor.

Constructors

ListSubscriptionsByTopicResponse'

Fields

nextToken :: Maybe Text

Token to pass along to the next ListSubscriptionsByTopic request. This element is returned if there are more subscriptions to retrieve.

subscriptions :: Maybe [Subscription]

A list of subscriptions.

httpStatus :: Int

The response's http status code.

newListSubscriptionsByTopicResponse

Arguments

:: Int
$sel:httpStatus:ListSubscriptionsByTopicResponse'

-> ListSubscriptionsByTopicResponse

Create a value of ListSubscriptionsByTopicResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSubscriptionsByTopic', listSubscriptionsByTopicResponse_nextToken - Token to pass along to the next ListSubscriptionsByTopic request. This element is returned if there are more subscriptions to retrieve.

$sel:subscriptions:ListSubscriptionsByTopicResponse', listSubscriptionsByTopicResponse_subscriptions - A list of subscriptions.

$sel:httpStatus:ListSubscriptionsByTopicResponse', listSubscriptionsByTopicResponse_httpStatus - The response's http status code.

Response Lenses

listSubscriptionsByTopicResponse_nextToken :: Lens' ListSubscriptionsByTopicResponse (Maybe Text)

Token to pass along to the next ListSubscriptionsByTopic request. This element is returned if there are more subscriptions to retrieve.

listSubscriptionsByTopicResponse_subscriptions :: Lens' ListSubscriptionsByTopicResponse (Maybe [Subscription])

A list of subscriptions.

listSubscriptionsByTopicResponse_httpStatus :: Lens' ListSubscriptionsByTopicResponse Int

The response's http status code.