Network.AWS.SNS.Types

Description

Service Configuration

defaultService :: Service

API version 2010-03-31 of the Amazon Simple Notification Service SDK configuration.

Errors

_KMSInvalidStateException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the state of the specified resource isn't valid for this request. For more information, see How Key State Affects Use of a Customer Master Key in the Key Management Service Developer Guide.

_EndpointDisabledException :: AsError a => Getting (First ServiceError) a ServiceError

Exception error indicating endpoint disabled.

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that a parameter in the request is invalid.

_AuthorizationErrorException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the user has been denied access to the requested resource.

_KMSThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling. For more information about throttling, see Limits in the Key Management Service Developer Guide.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that a request parameter does not comply with the associated constraints.

_SubscriptionLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the customer already owns the maximum allowed number of subscriptions.

_PlatformApplicationDisabledException :: AsError a => Getting (First ServiceError) a ServiceError

Exception error indicating platform application disabled.

_KMSOptInRequired :: AsError a => Getting (First ServiceError) a ServiceError

The Amazon Web Services access key ID needs a subscription for the service.

_InternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates an internal service error.

_ThrottledException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the rate at which requests have been submitted for this action exceeds the limit for your account.

_KMSNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the specified entity or resource can't be found.

_InvalidParameterValueException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that a request parameter does not comply with the associated constraints.

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the requested resource does not exist.

_StaleTagException :: AsError a => Getting (First ServiceError) a ServiceError

A tag has been added to a resource with the same ARN as a deleted resource. Wait a short while and then retry the operation.

_KMSDisabledException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the specified customer master key (CMK) isn't enabled.

_TagPolicyException :: AsError a => Getting (First ServiceError) a ServiceError

The request doesn't comply with the IAM tag policy. Correct your request and then retry it.

_InvalidSecurityException :: AsError a => Getting (First ServiceError) a ServiceError

The credential signature isn't valid. You must use an HTTPS endpoint and sign your request using Signature Version 4.

_TopicLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the customer already owns the maximum allowed number of topics.

_ConcurrentAccessException :: AsError a => Getting (First ServiceError) a ServiceError

Can't perform multiple operations on a tag simultaneously. Perform the operations sequentially.

_OptedOutException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the specified phone number opted out of receiving SMS messages from your account. You can't send SMS messages to phone numbers that opt out.

_TagLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Can't add more than 50 tags to a topic.

_UserErrorException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that a request parameter does not comply with the associated constraints.

_VerificationException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the one-time password (OTP) used for verification is invalid.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Can’t perform the action on the specified resource. Make sure that the resource exists.

_FilterPolicyLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the number of filter polices in your account exceeds the limit. To add more filter polices, submit an SNS Limit Increase case in the Amazon Web Services Support Center.

_KMSAccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

The ciphertext references a key that doesn't exist or that you don't have access to.

LanguageCodeString

newtype LanguageCodeString

Supported language code for sending OTP message

Constructors

LanguageCodeString'

Fields

fromLanguageCodeString :: Text

Bundled Patterns

pattern LanguageCodeString_De_DE :: LanguageCodeString

pattern LanguageCodeString_En_GB :: LanguageCodeString

pattern LanguageCodeString_En_US :: LanguageCodeString

pattern LanguageCodeString_Es_419 :: LanguageCodeString

pattern LanguageCodeString_Es_ES :: LanguageCodeString

pattern LanguageCodeString_Fr_CA :: LanguageCodeString

pattern LanguageCodeString_Fr_FR :: LanguageCodeString

pattern LanguageCodeString_It_IT :: LanguageCodeString

pattern LanguageCodeString_Ja_JP :: LanguageCodeString

pattern LanguageCodeString_Kr_KR :: LanguageCodeString

pattern LanguageCodeString_Pt_BR :: LanguageCodeString

pattern LanguageCodeString_Zh_CN :: LanguageCodeString

pattern LanguageCodeString_Zh_TW :: LanguageCodeString

NumberCapability

newtype NumberCapability

Enum listing out all supported number capabilities.

Constructors

NumberCapability'

Fields

fromNumberCapability :: Text

Bundled Patterns

pattern NumberCapability_MMS :: NumberCapability

pattern NumberCapability_SMS :: NumberCapability

pattern NumberCapability_VOICE :: NumberCapability

RouteType

newtype RouteType

Enum listing out all supported route types. The following enum values are supported. 1. Transactional : Non-marketing traffic 2. Promotional : Marketing 3. Premium : Premium routes for OTP delivery to the carriers

Constructors

RouteType'

Fields

fromRouteType :: Text

Bundled Patterns

pattern RouteType_Premium :: RouteType

pattern RouteType_Promotional :: RouteType

pattern RouteType_Transactional :: RouteType

SMSSandboxPhoneNumberVerificationStatus

newtype SMSSandboxPhoneNumberVerificationStatus

Enum listing out all supported destination phone number verification statuses. The following enum values are supported. 1. PENDING : The destination phone number is pending verification. 2. VERIFIED : The destination phone number is verified.

Constructors

SMSSandboxPhoneNumberVerificationStatus'

Fields

fromSMSSandboxPhoneNumberVerificationStatus :: Text

Bundled Patterns

pattern SMSSandboxPhoneNumberVerificationStatus_Pending :: SMSSandboxPhoneNumberVerificationStatus

pattern SMSSandboxPhoneNumberVerificationStatus_Verified :: SMSSandboxPhoneNumberVerificationStatus

Endpoint

data Endpoint

Endpoint for mobile app and device.

See: newEndpoint smart constructor.

Constructors

Endpoint'

Fields

attributes :: Maybe (HashMap Text Text)

Attributes for endpoint.

endpointArn :: Maybe Text

EndpointArn for mobile app and device.

newEndpoint :: Endpoint

Create a value of Endpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:Endpoint', endpoint_attributes - Attributes for endpoint.

$sel:endpointArn:Endpoint', endpoint_endpointArn - EndpointArn for mobile app and device.

endpoint_attributes :: Lens' Endpoint (Maybe (HashMap Text Text))

Attributes for endpoint.

endpoint_endpointArn :: Lens' Endpoint (Maybe Text)

EndpointArn for mobile app and device.

MessageAttributeValue

data MessageAttributeValue

The user-specified message attribute value. For string data types, the value attribute has the same restrictions on the content as the message body. For more information, see Publish.

Name, type, and value must not be empty or null. In addition, the message body should not be empty or null. All parts of the message attribute, including name, type, and value, are included in the message size restriction, which is currently 256 KB (262,144 bytes). For more information, see Amazon SNS message attributes and Publishing to a mobile phone in the Amazon SNS Developer Guide.

See: newMessageAttributeValue smart constructor.

Constructors

MessageAttributeValue'

Fields

binaryValue :: Maybe Base64

Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.

stringValue :: Maybe Text

Strings are Unicode with UTF8 binary encoding. For a list of code values, see ASCII Printable Characters.

dataType :: Text

Amazon SNS supports the following logical data types: String, String.Array, Number, and Binary. For more information, see Message Attribute Data Types.

newMessageAttributeValue

Arguments

:: Text
$sel:dataType:MessageAttributeValue'

-> MessageAttributeValue

Create a value of MessageAttributeValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:binaryValue:MessageAttributeValue', messageAttributeValue_binaryValue - Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:stringValue:MessageAttributeValue', messageAttributeValue_stringValue - Strings are Unicode with UTF8 binary encoding. For a list of code values, see ASCII Printable Characters.

$sel:dataType:MessageAttributeValue', messageAttributeValue_dataType - Amazon SNS supports the following logical data types: String, String.Array, Number, and Binary. For more information, see Message Attribute Data Types.

messageAttributeValue_binaryValue :: Lens' MessageAttributeValue (Maybe ByteString)

Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

messageAttributeValue_stringValue :: Lens' MessageAttributeValue (Maybe Text)

Strings are Unicode with UTF8 binary encoding. For a list of code values, see ASCII Printable Characters.

messageAttributeValue_dataType :: Lens' MessageAttributeValue Text

Amazon SNS supports the following logical data types: String, String.Array, Number, and Binary. For more information, see Message Attribute Data Types.

PhoneNumberInformation

data PhoneNumberInformation

A list of phone numbers and their metadata.

See: newPhoneNumberInformation smart constructor.

Constructors

PhoneNumberInformation'

Fields

status :: Maybe Text

The status of the phone number.

iso2CountryCode :: Maybe Text

The two-character code for the country or region, in ISO 3166-1 alpha-2 format.

createdAt :: Maybe ISO8601

The date and time when the phone number was created.

phoneNumber :: Maybe Text

The phone number.

numberCapabilities :: Maybe [NumberCapability]

The capabilities of each phone number.

routeType :: Maybe RouteType

The list of supported routes.

newPhoneNumberInformation :: PhoneNumberInformation

Create a value of PhoneNumberInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:PhoneNumberInformation', phoneNumberInformation_status - The status of the phone number.

$sel:iso2CountryCode:PhoneNumberInformation', phoneNumberInformation_iso2CountryCode - The two-character code for the country or region, in ISO 3166-1 alpha-2 format.

$sel:createdAt:PhoneNumberInformation', phoneNumberInformation_createdAt - The date and time when the phone number was created.

$sel:phoneNumber:PhoneNumberInformation', phoneNumberInformation_phoneNumber - The phone number.

$sel:numberCapabilities:PhoneNumberInformation', phoneNumberInformation_numberCapabilities - The capabilities of each phone number.

$sel:routeType:PhoneNumberInformation', phoneNumberInformation_routeType - The list of supported routes.

phoneNumberInformation_status :: Lens' PhoneNumberInformation (Maybe Text)

The status of the phone number.

phoneNumberInformation_iso2CountryCode :: Lens' PhoneNumberInformation (Maybe Text)

The two-character code for the country or region, in ISO 3166-1 alpha-2 format.

phoneNumberInformation_createdAt :: Lens' PhoneNumberInformation (Maybe UTCTime)

The date and time when the phone number was created.

phoneNumberInformation_phoneNumber :: Lens' PhoneNumberInformation (Maybe Text)

The phone number.

phoneNumberInformation_numberCapabilities :: Lens' PhoneNumberInformation (Maybe [NumberCapability])

The capabilities of each phone number.

phoneNumberInformation_routeType :: Lens' PhoneNumberInformation (Maybe RouteType)

The list of supported routes.

PlatformApplication

data PlatformApplication

Platform application object.

See: newPlatformApplication smart constructor.

Constructors

PlatformApplication'

Fields

platformApplicationArn :: Maybe Text

PlatformApplicationArn for platform application object.

attributes :: Maybe (HashMap Text Text)

Attributes for platform application object.

newPlatformApplication :: PlatformApplication

Create a value of PlatformApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platformApplicationArn:PlatformApplication', platformApplication_platformApplicationArn - PlatformApplicationArn for platform application object.

$sel:attributes:PlatformApplication', platformApplication_attributes - Attributes for platform application object.

platformApplication_platformApplicationArn :: Lens' PlatformApplication (Maybe Text)

PlatformApplicationArn for platform application object.

platformApplication_attributes :: Lens' PlatformApplication (Maybe (HashMap Text Text))

Attributes for platform application object.

SMSSandboxPhoneNumber

data SMSSandboxPhoneNumber

A verified or pending destination phone number in the SMS sandbox.

When you start using Amazon SNS to send SMS messages, your account is in the SMS sandbox. The SMS sandbox provides a safe environment for you to try Amazon SNS features without risking your reputation as an SMS sender. While your account is in the SMS sandbox, you can use all of the features of Amazon SNS. However, you can send SMS messages only to verified destination phone numbers. For more information, including how to move out of the sandbox to send messages without restrictions, see SMS sandbox in the Amazon SNS Developer Guide.

See: newSMSSandboxPhoneNumber smart constructor.

Constructors

SMSSandboxPhoneNumber'

Fields

status :: Maybe SMSSandboxPhoneNumberVerificationStatus

The destination phone number's verification status.

phoneNumber :: Maybe Text

The destination phone number.

newSMSSandboxPhoneNumber :: SMSSandboxPhoneNumber

Create a value of SMSSandboxPhoneNumber with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SMSSandboxPhoneNumber', sMSSandboxPhoneNumber_status - The destination phone number's verification status.

$sel:phoneNumber:SMSSandboxPhoneNumber', sMSSandboxPhoneNumber_phoneNumber - The destination phone number.

sMSSandboxPhoneNumber_status :: Lens' SMSSandboxPhoneNumber (Maybe SMSSandboxPhoneNumberVerificationStatus)

The destination phone number's verification status.

sMSSandboxPhoneNumber_phoneNumber :: Lens' SMSSandboxPhoneNumber (Maybe Text)

The destination phone number.

Subscription

data Subscription

A wrapper type for the attributes of an Amazon SNS subscription.

See: newSubscription smart constructor.

Constructors

Subscription'

Fields

protocol :: Maybe Text

The subscription's protocol.

owner :: Maybe Text

The subscription's owner.

topicArn :: Maybe Text

The ARN of the subscription's topic.

endpoint :: Maybe Text

The subscription's endpoint (format depends on the protocol).

subscriptionArn :: Maybe Text

The subscription's ARN.

newSubscription :: Subscription

Create a value of Subscription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocol:Subscription', subscription_protocol - The subscription's protocol.

$sel:owner:Subscription', subscription_owner - The subscription's owner.

$sel:topicArn:Subscription', subscription_topicArn - The ARN of the subscription's topic.

$sel:endpoint:Subscription', subscription_endpoint - The subscription's endpoint (format depends on the protocol).

$sel:subscriptionArn:Subscription', subscription_subscriptionArn - The subscription's ARN.

subscription_protocol :: Lens' Subscription (Maybe Text)

The subscription's protocol.

subscription_owner :: Lens' Subscription (Maybe Text)

The subscription's owner.

subscription_topicArn :: Lens' Subscription (Maybe Text)

The ARN of the subscription's topic.

subscription_endpoint :: Lens' Subscription (Maybe Text)

The subscription's endpoint (format depends on the protocol).

subscription_subscriptionArn :: Lens' Subscription (Maybe Text)

The subscription's ARN.

Tag

data Tag

The list of tags to be added to the specified topic.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

The required key portion of the tag.

value :: Text

The optional value portion of the tag.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The required key portion of the tag.

$sel:value:Tag', tag_value - The optional value portion of the tag.

tag_key :: Lens' Tag Text

The required key portion of the tag.

tag_value :: Lens' Tag Text

The optional value portion of the tag.

Topic

data Topic

A wrapper type for the topic's Amazon Resource Name (ARN). To retrieve a topic's attributes, use GetTopicAttributes.

See: newTopic smart constructor.

Constructors

Topic'

Fields

topicArn :: Maybe Text

The topic's ARN.

newTopic :: Topic

Create a value of Topic with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:topicArn:Topic', topic_topicArn - The topic's ARN.

topic_topicArn :: Lens' Topic (Maybe Text)

The topic's ARN.