Network.AWS.SNS.Types.Endpoint

Description

Documentation

data Endpoint

Endpoint for mobile app and device.

See: newEndpoint smart constructor.

Constructors

Endpoint'

Fields

attributes :: Maybe (HashMap Text Text)

Attributes for endpoint.

endpointArn :: Maybe Text

EndpointArn for mobile app and device.

newEndpoint :: Endpoint

Create a value of Endpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:Endpoint', endpoint_attributes - Attributes for endpoint.

$sel:endpointArn:Endpoint', endpoint_endpointArn - EndpointArn for mobile app and device.

endpoint_attributes :: Lens' Endpoint (Maybe (HashMap Text Text))

Attributes for endpoint.

endpoint_endpointArn :: Lens' Endpoint (Maybe Text)

EndpointArn for mobile app and device.