Network.AWS.SNS.Types.MessageAttributeValue

Description

Documentation

data MessageAttributeValue

The user-specified message attribute value. For string data types, the value attribute has the same restrictions on the content as the message body. For more information, see Publish.

Name, type, and value must not be empty or null. In addition, the message body should not be empty or null. All parts of the message attribute, including name, type, and value, are included in the message size restriction, which is currently 256 KB (262,144 bytes). For more information, see Amazon SNS message attributes and Publishing to a mobile phone in the Amazon SNS Developer Guide.

See: newMessageAttributeValue smart constructor.

Constructors

MessageAttributeValue'

Fields

stringValue :: Maybe Text

Strings are Unicode with UTF8 binary encoding. For a list of code values, see ASCII Printable Characters.

binaryValue :: Maybe Base64

Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.

dataType :: Text

Amazon SNS supports the following logical data types: String, String.Array, Number, and Binary. For more information, see Message Attribute Data Types.

newMessageAttributeValue

Arguments

:: Text
$sel:dataType:MessageAttributeValue'

-> MessageAttributeValue

Create a value of MessageAttributeValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stringValue:MessageAttributeValue', messageAttributeValue_stringValue - Strings are Unicode with UTF8 binary encoding. For a list of code values, see ASCII Printable Characters.

$sel:binaryValue:MessageAttributeValue', messageAttributeValue_binaryValue - Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:dataType:MessageAttributeValue', messageAttributeValue_dataType - Amazon SNS supports the following logical data types: String, String.Array, Number, and Binary. For more information, see Message Attribute Data Types.

messageAttributeValue_stringValue :: Lens' MessageAttributeValue (Maybe Text)

Strings are Unicode with UTF8 binary encoding. For a list of code values, see ASCII Printable Characters.

messageAttributeValue_binaryValue :: Lens' MessageAttributeValue (Maybe ByteString)

Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

messageAttributeValue_dataType :: Lens' MessageAttributeValue Text

Amazon SNS supports the following logical data types: String, String.Array, Number, and Binary. For more information, see Message Attribute Data Types.