Network.AWS.SNS.Types.PhoneNumberInformation

Description

Documentation

data PhoneNumberInformation

A list of phone numbers and their metadata.

See: newPhoneNumberInformation smart constructor.

Constructors

PhoneNumberInformation'

Fields

status :: Maybe Text

The status of the phone number.

iso2CountryCode :: Maybe Text

The two-character code for the country or region, in ISO 3166-1 alpha-2 format.

createdAt :: Maybe ISO8601

The date and time when the phone number was created.

phoneNumber :: Maybe Text

The phone number.

numberCapabilities :: Maybe [NumberCapability]

The capabilities of each phone number.

routeType :: Maybe RouteType

The list of supported routes.

newPhoneNumberInformation :: PhoneNumberInformation

Create a value of PhoneNumberInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:PhoneNumberInformation', phoneNumberInformation_status - The status of the phone number.

$sel:iso2CountryCode:PhoneNumberInformation', phoneNumberInformation_iso2CountryCode - The two-character code for the country or region, in ISO 3166-1 alpha-2 format.

$sel:createdAt:PhoneNumberInformation', phoneNumberInformation_createdAt - The date and time when the phone number was created.

$sel:phoneNumber:PhoneNumberInformation', phoneNumberInformation_phoneNumber - The phone number.

$sel:numberCapabilities:PhoneNumberInformation', phoneNumberInformation_numberCapabilities - The capabilities of each phone number.

$sel:routeType:PhoneNumberInformation', phoneNumberInformation_routeType - The list of supported routes.

phoneNumberInformation_status :: Lens' PhoneNumberInformation (Maybe Text)

The status of the phone number.

phoneNumberInformation_iso2CountryCode :: Lens' PhoneNumberInformation (Maybe Text)

The two-character code for the country or region, in ISO 3166-1 alpha-2 format.

phoneNumberInformation_createdAt :: Lens' PhoneNumberInformation (Maybe UTCTime)

The date and time when the phone number was created.

phoneNumberInformation_phoneNumber :: Lens' PhoneNumberInformation (Maybe Text)

The phone number.

phoneNumberInformation_numberCapabilities :: Lens' PhoneNumberInformation (Maybe [NumberCapability])

The capabilities of each phone number.

phoneNumberInformation_routeType :: Lens' PhoneNumberInformation (Maybe RouteType)

The list of supported routes.