Network.AWS.SNS.V2010_03_31.ConfirmSubscription

Description

Verifies an endpoint owner's intent to receive messages by validating the token sent to the endpoint by an earlier Subscribe action. If the token is valid, the action creates a new subscription and returns its Amazon Resource Name (ARN). This call requires an AWS signature only when the AuthenticateOnUnsubscribe flag is set to "true". https:/sns.us-east-1.amazonaws.com ?Action=ConfirmSubscription &amp;TopicArn=arn:aws:sns:us-east-1:123456789012:My-Topic &amp;Token=51b2ff3edb475b7d91550e0ab6edf0c1de2a34e6ebaf6 c2262a001bcb7e051c43aa00022ceecce70bd2a67b2042da8d8 eb47fef7a4e4e942d23e7fa56146b9ee35da040b4b8af564cc4 184a7391c834cb75d75c22981f776ad1ce8805e9bab29da2329 985337bb8095627907b46c8577c8440556b6f86582a95475802 6f41fc62041c4b3f67b0f5921232b5dae5aaca1 &lt;ConfirmSubscriptionResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/"&gt; &lt;ConfirmSubscriptionResult&gt; &lt;SubscriptionArn&gt;arn:aws:sns:us-east-1:123456789012:My-Topic:80289ba6-0fd4-4079-afb4-ce8c8260f0ca&lt;/SubscriptionArn&gt; &lt;/ConfirmSubscriptionResult&gt; &lt;ResponseMetadata&gt; &lt;RequestId&gt;7a50221f-3774-11df-a9b7-05d48da6f042&lt;/RequestId&gt; &lt;ResponseMetadata&gt; &lt;ConfirmSubscriptionResponse&gt;.

Documentation

confirmSubscription

Arguments

:: Text
_csiToken

-> Text
_csiTopicArn

-> ConfirmSubscription

Minimum specification for a ConfirmSubscription request.

data ConfirmSubscription

Constructors

ConfirmSubscription

Fields

_csiToken :: Text

Short-lived token sent to an endpoint during the Subscribe action.

_csiTopicArn :: Text

The ARN of the topic for which you wish to confirm a subscription.

_csiAuthenticateOnUnsubscribe :: Maybe Text

Disallows unauthenticated unsubscribes of the subscription. If the value of this parameter is true and the request has an AWS signature, then only the topic owner and the subscription owner can unsubscribe the endpoint. The unsubscribe action requires AWS authentication.

csiTopicArn :: Lens' ConfirmSubscription Text

csiToken :: Lens' ConfirmSubscription Text

csiAuthenticateOnUnsubscribe :: Lens' ConfirmSubscription (Maybe Text)

data ConfirmSubscriptionResponse

Constructors

ConfirmSubscriptionResponse

Fields

_csrSubscriptionArn :: Maybe Text

The ARN of the created subscription.

csrSubscriptionArn :: Iso' ConfirmSubscriptionResponse (Maybe Text)