Network.AWS.SNS.V2010_03_31.Types

Description

Amazon Simple Notification Service (Amazon SNS) is a fast, flexible, fully managed push messaging service. Amazon SNS makes it simple and cost-effective to push notifications to Apple, Google, Fire OS, and Windows devices, as well as Android devices in China with Baidu Cloud Push. You can also use SNS to push notifications to internet connected smart devices, as well as other distributed services. Besides pushing cloud notifications directly to mobile devices, Amazon SNS can also deliver notifications by SMS text message or email, to Amazon Simple Queue Service (SQS) queues, or to any HTTP endpoint. To prevent messages from being lost, all messages published to Amazon SNS are stored redundantly across multiple availability zones.

Documentation

data SNS

Supported version (2010-03-31) of the Amazon Simple Notification Service service.

xmlOptions :: Tagged a XMLOptions

newtype Topic

A wrapper type for the topic's Amazon Resource Name (ARN). To retrieve a topic's attributes, use GetTopicAttributes.

Constructors

Topic

Fields

_tTopicArn :: Maybe Text

The topic's ARN.

data Endpoint

Endpoint for mobile app and device.

Constructors

Endpoint

Fields

_fAttributes :: HashMap Text Text

Attributes for endpoint.

_fEndpointArn :: Maybe Text

EndpointArn for mobile app and device.

data MessageAttributeValue

The user-specified message attribute value. For string data types, the value attribute has the same restrictions on the content as the message body. For more information, see Publish. Name, type, and value must not be empty or null. In addition, the message body should not be empty or null. All parts of the message attribute, including name, type, and value, are included in the message size restriction, which is currently 256 KB (262,144 bytes). For more information, see Using Amazon SNS Message Attributes.

Constructors

MessageAttributeValue

Fields

_mavBinaryValue :: Maybe ByteString

Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.

_mavStringValue :: Maybe Text

Strings are Unicode with UTF8 binary encoding. For a list of code values, see http:/en.wikipedia.orgwiki/ASCII#ASCII_printable_characters.

_mavDataType :: Text

Amazon SNS supports the following logical data types: String, Number, and Binary. For more information, see Message Attribute Data Types.

data PlatformApplication

Platform application object.

Constructors

PlatformApplication

Fields

_paPlatformApplicationArn :: Maybe Text

PlatformApplicationArn for platform application object.

_paAttributes :: HashMap Text Text

Attributes for platform application object.

data Subscription

A wrapper type for the attributes of an Amazon SNS subscription.

Constructors

Subscription

Fields

_snProtocol :: Maybe Text

The subscription's protocol.

_snOwner :: Maybe Text

The subscription's owner.

_snTopicArn :: Maybe Text

The ARN of the subscription's topic.

_snEndpoint :: Maybe Text

The subscription's endpoint (format depends on the protocol).

_snSubscriptionArn :: Maybe Text

The subscription's ARN.

tTopicArn :: Iso' Topic (Maybe Text)

topic :: Iso' (Maybe Text) Topic

fEndpointArn :: Lens' Endpoint (Maybe Text)

fAttributes :: Lens' Endpoint (HashMap Text Text)

mavStringValue :: Lens' MessageAttributeValue (Maybe Text)

mavDataType :: Lens' MessageAttributeValue Text

mavBinaryValue :: Lens' MessageAttributeValue (Maybe ByteString)

paPlatformApplicationArn :: Lens' PlatformApplication (Maybe Text)

paAttributes :: Lens' PlatformApplication (HashMap Text Text)

snTopicArn :: Lens' Subscription (Maybe Text)

snSubscriptionArn :: Lens' Subscription (Maybe Text)

snProtocol :: Lens' Subscription (Maybe Text)

snOwner :: Lens' Subscription (Maybe Text)

snEndpoint :: Lens' Subscription (Maybe Text)