Network.AWS.SQS.ChangeMessageVisibilityBatch

Description

Changes the visibility timeout of multiple messages. This is a batch version of ChangeMessageVisibility. The result of the action on each message is reported individually in the response. You can send up to 10 ChangeMessageVisibility requests with each ChangeMessageVisibilityBatch action.

Because the batch request can result in a combination of successful and unsuccessful actions, you should check for batch errors even when the call returns an HTTP status code of 200.

Some actions take lists of parameters. These lists are specified using the param.n notation. Values of n are integers starting from 1. For example, a parameter list with two elements looks like this:

&AttributeName.1=first

&AttributeName.2=second

Creating a Request

data ChangeMessageVisibilityBatch

See: newChangeMessageVisibilityBatch smart constructor.

Constructors

ChangeMessageVisibilityBatch'

Fields

queueUrl :: Text

The URL of the Amazon SQS queue whose messages' visibility is changed.

Queue URLs and names are case-sensitive.

entries :: [ChangeMessageVisibilityBatchRequestEntry]

A list of receipt handles of the messages for which the visibility timeout must be changed.

newChangeMessageVisibilityBatch

Arguments

:: Text
$sel:queueUrl:ChangeMessageVisibilityBatch'

-> ChangeMessageVisibilityBatch

Create a value of ChangeMessageVisibilityBatch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queueUrl:ChangeMessageVisibilityBatch', changeMessageVisibilityBatch_queueUrl - The URL of the Amazon SQS queue whose messages' visibility is changed.

Queue URLs and names are case-sensitive.

$sel:entries:ChangeMessageVisibilityBatch', changeMessageVisibilityBatch_entries - A list of receipt handles of the messages for which the visibility timeout must be changed.

Request Lenses

changeMessageVisibilityBatch_queueUrl :: Lens' ChangeMessageVisibilityBatch Text

The URL of the Amazon SQS queue whose messages' visibility is changed.

Queue URLs and names are case-sensitive.

changeMessageVisibilityBatch_entries :: Lens' ChangeMessageVisibilityBatch [ChangeMessageVisibilityBatchRequestEntry]

A list of receipt handles of the messages for which the visibility timeout must be changed.

Destructuring the Response

data ChangeMessageVisibilityBatchResponse

For each message in the batch, the response contains a ChangeMessageVisibilityBatchResultEntry tag if the message succeeds or a BatchResultErrorEntry tag if the message fails.

See: newChangeMessageVisibilityBatchResponse smart constructor.

Constructors

ChangeMessageVisibilityBatchResponse'

Fields

httpStatus :: Int

The response's http status code.

successful :: [ChangeMessageVisibilityBatchResultEntry]

A list of ChangeMessageVisibilityBatchResultEntry items.

failed :: [BatchResultErrorEntry]

A list of BatchResultErrorEntry items.

newChangeMessageVisibilityBatchResponse

Arguments

:: Int
$sel:httpStatus:ChangeMessageVisibilityBatchResponse'

-> ChangeMessageVisibilityBatchResponse

Create a value of ChangeMessageVisibilityBatchResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ChangeMessageVisibilityBatchResponse', changeMessageVisibilityBatchResponse_httpStatus - The response's http status code.

$sel:successful:ChangeMessageVisibilityBatchResponse', changeMessageVisibilityBatchResponse_successful - A list of ChangeMessageVisibilityBatchResultEntry items.

$sel:failed:ChangeMessageVisibilityBatchResponse', changeMessageVisibilityBatchResponse_failed - A list of BatchResultErrorEntry items.

Response Lenses

changeMessageVisibilityBatchResponse_httpStatus :: Lens' ChangeMessageVisibilityBatchResponse Int

The response's http status code.

changeMessageVisibilityBatchResponse_successful :: Lens' ChangeMessageVisibilityBatchResponse [ChangeMessageVisibilityBatchResultEntry]

A list of ChangeMessageVisibilityBatchResultEntry items.

changeMessageVisibilityBatchResponse_failed :: Lens' ChangeMessageVisibilityBatchResponse [BatchResultErrorEntry]

A list of BatchResultErrorEntry items.