Network.AWS.SQS.DeleteMessage

Description

Deletes the specified message from the specified queue. To select the message to delete, use the ReceiptHandle of the message (not the MessageId which you receive when you send the message). Amazon SQS can delete a message from a queue even if a visibility timeout setting causes the message to be locked by another consumer. Amazon SQS automatically deletes messages left in a queue longer than the retention period configured for the queue.

The ReceiptHandle is associated with a specific instance of receiving a message. If you receive a message more than once, the ReceiptHandle is different each time you receive a message. When you use the DeleteMessage action, you must provide the most recently received ReceiptHandle for the message (otherwise, the request succeeds, but the message might not be deleted).

For standard queues, it is possible to receive a message even after you delete it. This might happen on rare occasions if one of the servers which stores a copy of the message is unavailable when you send the request to delete the message. The copy remains on the server and might be returned to you during a subsequent receive request. You should ensure that your application is idempotent, so that receiving a message more than once does not cause issues.

Creating a Request

data DeleteMessage

See: newDeleteMessage smart constructor.

Constructors

DeleteMessage'

Fields

queueUrl :: Text

The URL of the Amazon SQS queue from which messages are deleted.

Queue URLs and names are case-sensitive.

receiptHandle :: Text

The receipt handle associated with the message to delete.

newDeleteMessage

Arguments

:: Text
$sel:queueUrl:DeleteMessage'

-> Text
$sel:receiptHandle:DeleteMessage'

-> DeleteMessage

Create a value of DeleteMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queueUrl:DeleteMessage', deleteMessage_queueUrl - The URL of the Amazon SQS queue from which messages are deleted.

Queue URLs and names are case-sensitive.

$sel:receiptHandle:DeleteMessage', deleteMessage_receiptHandle - The receipt handle associated with the message to delete.

Request Lenses

deleteMessage_queueUrl :: Lens' DeleteMessage Text

The URL of the Amazon SQS queue from which messages are deleted.

Queue URLs and names are case-sensitive.

deleteMessage_receiptHandle :: Lens' DeleteMessage Text

The receipt handle associated with the message to delete.

Destructuring the Response

data DeleteMessageResponse

See: newDeleteMessageResponse smart constructor.

Constructors

DeleteMessageResponse'

newDeleteMessageResponse :: DeleteMessageResponse

Create a value of DeleteMessageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.