Network.AWS.SQS.DeleteQueue

Description

Deletes the queue specified by the QueueUrl, regardless of the queue's contents.

Be careful with the DeleteQueue action: When you delete a queue, any messages in the queue are no longer available.

When you delete a queue, the deletion process takes up to 60 seconds. Requests you send involving that queue during the 60 seconds might succeed. For example, a SendMessage request might succeed, but after 60 seconds the queue and the message you sent no longer exist.

When you delete a queue, you must wait at least 60 seconds before creating a queue with the same name.

Cross-account permissions don't apply to this action. For more information, see Grant cross-account permissions to a role and a user name in the Amazon SQS Developer Guide.

Creating a Request

data DeleteQueue

See: newDeleteQueue smart constructor.

Constructors

DeleteQueue'

Fields

queueUrl :: Text

The URL of the Amazon SQS queue to delete.

Queue URLs and names are case-sensitive.

newDeleteQueue

Arguments

:: Text
$sel:queueUrl:DeleteQueue'

-> DeleteQueue

Create a value of DeleteQueue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queueUrl:DeleteQueue', deleteQueue_queueUrl - The URL of the Amazon SQS queue to delete.

Queue URLs and names are case-sensitive.

Request Lenses

deleteQueue_queueUrl :: Lens' DeleteQueue Text

The URL of the Amazon SQS queue to delete.

Queue URLs and names are case-sensitive.

Destructuring the Response

data DeleteQueueResponse

See: newDeleteQueueResponse smart constructor.

Constructors

DeleteQueueResponse'

newDeleteQueueResponse :: DeleteQueueResponse

Create a value of DeleteQueueResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.