Network.AWS.SQS.Lens

Description

Operations

GetQueueUrl

getQueueUrl_queueOwnerAWSAccountId :: Lens' GetQueueUrl (Maybe Text)

The account ID of the account that created the queue.

getQueueUrl_queueName :: Lens' GetQueueUrl Text

The name of the queue whose URL must be fetched. Maximum 80 characters. Valid values: alphanumeric characters, hyphens (-), and underscores (_).

Queue URLs and names are case-sensitive.

getQueueUrlResponse_httpStatus :: Lens' GetQueueUrlResponse Int

The response's http status code.

getQueueUrlResponse_queueUrl :: Lens' GetQueueUrlResponse Text

The URL of the queue.

PurgeQueue

purgeQueue_queueUrl :: Lens' PurgeQueue Text

The URL of the queue from which the PurgeQueue action deletes messages.

Queue URLs and names are case-sensitive.

ChangeMessageVisibilityBatch

changeMessageVisibilityBatch_queueUrl :: Lens' ChangeMessageVisibilityBatch Text

The URL of the Amazon SQS queue whose messages' visibility is changed.

Queue URLs and names are case-sensitive.

changeMessageVisibilityBatch_entries :: Lens' ChangeMessageVisibilityBatch [ChangeMessageVisibilityBatchRequestEntry]

A list of receipt handles of the messages for which the visibility timeout must be changed.

changeMessageVisibilityBatchResponse_httpStatus :: Lens' ChangeMessageVisibilityBatchResponse Int

The response's http status code.

changeMessageVisibilityBatchResponse_successful :: Lens' ChangeMessageVisibilityBatchResponse [ChangeMessageVisibilityBatchResultEntry]

A list of ChangeMessageVisibilityBatchResultEntry items.

changeMessageVisibilityBatchResponse_failed :: Lens' ChangeMessageVisibilityBatchResponse [BatchResultErrorEntry]

A list of BatchResultErrorEntry items.

SendMessage

sendMessage_messageAttributes :: Lens' SendMessage (Maybe (HashMap Text MessageAttributeValue))

Each message attribute consists of a Name, Type, and Value. For more information, see Amazon SQS message attributes in the Amazon SQS Developer Guide.

sendMessage_delaySeconds :: Lens' SendMessage (Maybe Int)

The length of time, in seconds, for which to delay a specific message. Valid values: 0 to 900. Maximum: 15 minutes. Messages with a positive DelaySeconds value become available for processing after the delay period is finished. If you don't specify a value, the default value for the queue applies.

When you set FifoQueue, you can't set DelaySeconds per message. You can set this parameter only on a queue level.

sendMessage_messageSystemAttributes :: Lens' SendMessage (Maybe (HashMap MessageSystemAttributeNameForSends MessageSystemAttributeValue))

The message system attribute to send. Each message system attribute consists of a Name, Type, and Value.

Currently, the only supported message system attribute is AWSTraceHeader. Its type must be String and its value must be a correctly formatted X-Ray trace header string.

The size of a message system attribute doesn't count towards the total size of a message.

sendMessage_messageDeduplicationId :: Lens' SendMessage (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) queues.

The token used for deduplication of sent messages. If a message with a particular MessageDeduplicationId is sent successfully, any messages sent with the same MessageDeduplicationId are accepted successfully but aren't delivered during the 5-minute deduplication interval. For more information, see Exactly-once processing in the Amazon SQS Developer Guide.

Every message must have a unique MessageDeduplicationId,

You may provide a MessageDeduplicationId explicitly.

If you aren't able to provide a MessageDeduplicationId and you enable ContentBasedDeduplication for your queue, Amazon SQS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message).

If you don't provide a MessageDeduplicationId and the queue doesn't have ContentBasedDeduplication set, the action fails with an error.

If the queue has ContentBasedDeduplication set, your MessageDeduplicationId overrides the generated one.

When ContentBasedDeduplication is in effect, messages with identical content sent within the deduplication interval are treated as duplicates and only one copy of the message is delivered.

If you send one message with ContentBasedDeduplication enabled and then another message with a MessageDeduplicationId that is the same as the one generated for the first MessageDeduplicationId, the two messages are treated as duplicates and only one copy of the message is delivered.

The MessageDeduplicationId is available to the consumer of the message (this can be useful for troubleshooting delivery issues).

If a message is sent successfully but the acknowledgement is lost and the message is resent with the same MessageDeduplicationId after the deduplication interval, Amazon SQS can't detect duplicate messages.

Amazon SQS continues to keep track of the message deduplication ID even after the message is received and deleted.

The maximum length of MessageDeduplicationId is 128 characters. MessageDeduplicationId can contain alphanumeric characters (a-z, A-Z, 0-9) and punctuation (!"#$%&'()*+,-./:;<=>?@[\]^_`{|}~).

For best practices of using MessageDeduplicationId, see Using the MessageDeduplicationId Property in the Amazon SQS Developer Guide.

sendMessage_messageGroupId :: Lens' SendMessage (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) queues.

The tag that specifies that a message belongs to a specific message group. Messages that belong to the same message group are processed in a FIFO manner (however, messages in different message groups might be processed out of order). To interleave multiple ordered streams within a single queue, use MessageGroupId values (for example, session data for multiple users). In this scenario, multiple consumers can process the queue, but the session data of each user is processed in a FIFO fashion.

You must associate a non-empty MessageGroupId with a message. If you don't provide a MessageGroupId, the action fails.

ReceiveMessage might return messages with multiple MessageGroupId values. For each MessageGroupId, the messages are sorted by time sent. The caller can't specify a MessageGroupId.

The length of MessageGroupId is 128 characters. Valid values: alphanumeric characters and punctuation (!"#$%&'()*+,-./:;<=>?@[\]^_`{|}~).

For best practices of using MessageGroupId, see Using the MessageGroupId Property in the Amazon SQS Developer Guide.

MessageGroupId is required for FIFO queues. You can't use it for Standard queues.

sendMessage_queueUrl :: Lens' SendMessage Text

The URL of the Amazon SQS queue to which a message is sent.

Queue URLs and names are case-sensitive.

sendMessage_messageBody :: Lens' SendMessage Text

The message to send. The minimum size is one character. The maximum size is 256 KB.

A message can include only XML, JSON, and unformatted text. The following Unicode characters are allowed:

#x9 | #xA | #xD | #x20 to #xD7FF | #xE000 to #xFFFD | #x10000 to #x10FFFF

Any characters not included in this list will be rejected. For more information, see the W3C specification for characters.

sendMessageResponse_sequenceNumber :: Lens' SendMessageResponse (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) queues.

The large, non-consecutive number that Amazon SQS assigns to each message.

The length of SequenceNumber is 128 bits. SequenceNumber continues to increase for a particular MessageGroupId.

sendMessageResponse_mD5OfMessageSystemAttributes :: Lens' SendMessageResponse (Maybe Text)

An MD5 digest of the non-URL-encoded message system attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest.

sendMessageResponse_messageId :: Lens' SendMessageResponse (Maybe Text)

An attribute containing the MessageId of the message sent to the queue. For more information, see Queue and Message Identifiers in the Amazon SQS Developer Guide.

sendMessageResponse_mD5OfMessageBody :: Lens' SendMessageResponse (Maybe Text)

An MD5 digest of the non-URL-encoded message body string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

sendMessageResponse_mD5OfMessageAttributes :: Lens' SendMessageResponse (Maybe Text)

An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

sendMessageResponse_httpStatus :: Lens' SendMessageResponse Int

The response's http status code.

RemovePermission

removePermission_queueUrl :: Lens' RemovePermission Text

The URL of the Amazon SQS queue from which permissions are removed.

Queue URLs and names are case-sensitive.

removePermission_label :: Lens' RemovePermission Text

The identification of the permission to remove. This is the label added using the AddPermission action.

GetQueueAttributes

getQueueAttributes_attributeNames :: Lens' GetQueueAttributes (Maybe [QueueAttributeName])

A list of attributes for which to retrieve information.

The AttributeName.N parameter is optional, but if you don't specify values for this parameter, the request returns empty results.

In the future, new attributes might be added. If you write code that calls this action, we recommend that you structure your code so that it can handle new attributes gracefully.

The following attributes are supported:

The ApproximateNumberOfMessagesDelayed, ApproximateNumberOfMessagesNotVisible, and ApproximateNumberOfMessagesVisible metrics may not achieve consistency until at least 1 minute after the producers stop sending messages. This period is required for the queue metadata to reach eventual consistency.

All – Returns all values.

ApproximateNumberOfMessages – Returns the approximate number of messages available for retrieval from the queue.

ApproximateNumberOfMessagesDelayed – Returns the approximate number of messages in the queue that are delayed and not available for reading immediately. This can happen when the queue is configured as a delay queue or when a message has been sent with a delay parameter.

ApproximateNumberOfMessagesNotVisible – Returns the approximate number of messages that are in flight. Messages are considered to be in flight if they have been sent to a client but have not yet been deleted or have not yet reached the end of their visibility window.

CreatedTimestamp – Returns the time when the queue was created in seconds (epoch time).

DelaySeconds – Returns the default delay on the queue in seconds.

LastModifiedTimestamp – Returns the time when the queue was last changed in seconds (epoch time).

MaximumMessageSize – Returns the limit of how many bytes a message can contain before Amazon SQS rejects it.

MessageRetentionPeriod – Returns the length of time, in seconds, for which Amazon SQS retains a message.

Policy – Returns the policy of the queue.

QueueArn – Returns the Amazon resource name (ARN) of the queue.

ReceiveMessageWaitTimeSeconds – Returns the length of time, in seconds, for which the ReceiveMessage action waits for a message to arrive.

VisibilityTimeout – Returns the visibility timeout for the queue. For more information about the visibility timeout, see Visibility Timeout in the Amazon SQS Developer Guide.

The following attributes apply only to dead-letter queues:

RedrivePolicy – The string that includes the parameters for the dead-letter queue functionality of the source queue as a JSON object. The parameters are as follows:

deadLetterTargetArn – The Amazon Resource Name (ARN) of the dead-letter queue to which Amazon SQS moves messages after the value of maxReceiveCount is exceeded.

maxReceiveCount – The number of times a message is delivered to the source queue before being moved to the dead-letter queue. When the ReceiveCount for a message exceeds the maxReceiveCount for a queue, Amazon SQS moves the message to the dead-letter-queue.

RedriveAllowPolicy – The string that includes the parameters for the permissions for the dead-letter queue redrive permission and which source queues can specify dead-letter queues as a JSON object. The parameters are as follows:

redrivePermission – The permission type that defines which source queues can specify the current queue as the dead-letter queue. Valid values are:

allowAll – (Default) Any source queues in this Amazon Web Services account in the same Region can specify this queue as the dead-letter queue.

denyAll – No source queues can specify this queue as the dead-letter queue.

byQueue – Only queues specified by the sourceQueueArns parameter can specify this queue as the dead-letter queue.

sourceQueueArns – The Amazon Resource Names (ARN)s of the source queues that can specify this queue as the dead-letter queue and redrive messages. You can specify this parameter only when the redrivePermission parameter is set to byQueue. You can specify up to 10 source queue ARNs. To allow more than 10 source queues to specify dead-letter queues, set the redrivePermission parameter to allowAll.

The dead-letter queue of a FIFO queue must also be a FIFO queue. Similarly, the dead-letter queue of a standard queue must also be a standard queue.

The following attributes apply only to server-side-encryption:

KmsMasterKeyId – Returns the ID of an Amazon Web Services managed customer master key (CMK) for Amazon SQS or a custom CMK. For more information, see Key Terms.

KmsDataKeyReusePeriodSeconds – Returns the length of time, in seconds, for which Amazon SQS can reuse a data key to encrypt or decrypt messages before calling KMS again. For more information, see How Does the Data Key Reuse Period Work?.

The following attributes apply only to FIFO (first-in-first-out) queues:

FifoQueue – Returns information about whether the queue is FIFO. For more information, see FIFO queue logic in the Amazon SQS Developer Guide.

To determine whether a queue is FIFO, you can check whether QueueName ends with the .fifo suffix.

ContentBasedDeduplication – Returns whether content-based deduplication is enabled for the queue. For more information, see Exactly-once processing in the Amazon SQS Developer Guide.

The following attributes apply only to high throughput for FIFO queues:

DeduplicationScope – Specifies whether message deduplication occurs at the message group or queue level. Valid values are messageGroup and queue.

FifoThroughputLimit – Specifies whether the FIFO queue throughput quota applies to the entire queue or per message group. Valid values are perQueue and perMessageGroupId. The perMessageGroupId value is allowed only when the value for DeduplicationScope is messageGroup.

To enable high throughput for FIFO queues, do the following:

Set DeduplicationScope to messageGroup.

Set FifoThroughputLimit to perMessageGroupId.

If you set these attributes to anything other than the values shown for enabling high throughput, normal throughput is in effect and deduplication occurs as specified.

For information on throughput quotas, see Quotas related to messages in the Amazon SQS Developer Guide.

getQueueAttributes_queueUrl :: Lens' GetQueueAttributes Text

The URL of the Amazon SQS queue whose attribute information is retrieved.

Queue URLs and names are case-sensitive.

getQueueAttributesResponse_attributes :: Lens' GetQueueAttributesResponse (Maybe (HashMap QueueAttributeName Text))

A map of attributes to their respective values.

getQueueAttributesResponse_httpStatus :: Lens' GetQueueAttributesResponse Int

The response's http status code.

ListQueues

listQueues_queueNamePrefix :: Lens' ListQueues (Maybe Text)

A string to use for filtering the list results. Only those queues whose name begins with the specified string are returned.

Queue URLs and names are case-sensitive.

listQueues_nextToken :: Lens' ListQueues (Maybe Text)

Pagination token to request the next set of results.

listQueues_maxResults :: Lens' ListQueues (Maybe Int)

Maximum number of results to include in the response. Value range is 1 to 1000. You must set MaxResults to receive a value for NextToken in the response.

listQueuesResponse_queueUrls :: Lens' ListQueuesResponse (Maybe [Text])

A list of queue URLs, up to 1,000 entries, or the value of MaxResults that you sent in the request.

listQueuesResponse_nextToken :: Lens' ListQueuesResponse (Maybe Text)

Pagination token to include in the next request. Token value is null if there are no additional results to request, or if you did not set MaxResults in the request.

listQueuesResponse_httpStatus :: Lens' ListQueuesResponse Int

The response's http status code.

ReceiveMessage

receiveMessage_receiveRequestAttemptId :: Lens' ReceiveMessage (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) queues.

The token used for deduplication of ReceiveMessage calls. If a networking issue occurs after a ReceiveMessage action, and instead of a response you receive a generic error, it is possible to retry the same action with an identical ReceiveRequestAttemptId to retrieve the same set of messages, even if their visibility timeout has not yet expired.

You can use ReceiveRequestAttemptId only for 5 minutes after a ReceiveMessage action.

When you set FifoQueue, a caller of the ReceiveMessage action can provide a ReceiveRequestAttemptId explicitly.

If a caller of the ReceiveMessage action doesn't provide a ReceiveRequestAttemptId, Amazon SQS generates a ReceiveRequestAttemptId.

It is possible to retry the ReceiveMessage action with the same ReceiveRequestAttemptId if none of the messages have been modified (deleted or had their visibility changes).

During a visibility timeout, subsequent calls with the same ReceiveRequestAttemptId return the same messages and receipt handles. If a retry occurs within the deduplication interval, it resets the visibility timeout. For more information, see Visibility Timeout in the Amazon SQS Developer Guide.

If a caller of the ReceiveMessage action still processes messages when the visibility timeout expires and messages become visible, another worker consuming from the same queue can receive the same messages and therefore process duplicates. Also, if a consumer whose message processing time is longer than the visibility timeout tries to delete the processed messages, the action fails with an error.

To mitigate this effect, ensure that your application observes a safe threshold before the visibility timeout expires and extend the visibility timeout as necessary.

While messages with a particular MessageGroupId are invisible, no more messages belonging to the same MessageGroupId are returned until the visibility timeout expires. You can still receive messages with another MessageGroupId as long as it is also visible.

If a caller of ReceiveMessage can't track the ReceiveRequestAttemptId, no retries work until the original visibility timeout expires. As a result, delays might occur but the messages in the queue remain in a strict order.

The maximum length of ReceiveRequestAttemptId is 128 characters. ReceiveRequestAttemptId can contain alphanumeric characters (a-z, A-Z, 0-9) and punctuation (!"#$%&'()*+,-./:;<=>?@[\]^_`{|}~).

For best practices of using ReceiveRequestAttemptId, see Using the ReceiveRequestAttemptId Request Parameter in the Amazon SQS Developer Guide.

receiveMessage_visibilityTimeout :: Lens' ReceiveMessage (Maybe Int)

The duration (in seconds) that the received messages are hidden from subsequent retrieve requests after being retrieved by a ReceiveMessage request.

receiveMessage_messageAttributeNames :: Lens' ReceiveMessage (Maybe [Text])

The name of the message attribute, where N is the index.

The name can contain alphanumeric characters and the underscore (_), hyphen (-), and period (.).

The name is case-sensitive and must be unique among all attribute names for the message.

The name must not start with AWS-reserved prefixes such as AWS. or Amazon. (or any casing variants).

The name must not start or end with a period (.), and it should not have periods in succession (..).

The name can be up to 256 characters long.

When using ReceiveMessage, you can send a list of attribute names to receive, or you can return all of the attributes by specifying All or .* in your request. You can also use all message attributes starting with a prefix, for example bar.*.

receiveMessage_waitTimeSeconds :: Lens' ReceiveMessage (Maybe Int)

The duration (in seconds) for which the call waits for a message to arrive in the queue before returning. If a message is available, the call returns sooner than WaitTimeSeconds. If no messages are available and the wait time expires, the call returns successfully with an empty list of messages.

To avoid HTTP errors, ensure that the HTTP response timeout for ReceiveMessage requests is longer than the WaitTimeSeconds parameter. For example, with the Java SDK, you can set HTTP transport settings using the NettyNioAsyncHttpClient for asynchronous clients, or the ApacheHttpClient for synchronous clients.

receiveMessage_attributeNames :: Lens' ReceiveMessage (Maybe [MessageAttribute])

A list of attributes that need to be returned along with each message. These attributes include:

All – Returns all values.

ApproximateFirstReceiveTimestamp – Returns the time the message was first received from the queue (epoch time in milliseconds).

ApproximateReceiveCount – Returns the number of times a message has been received across all queues but not deleted.

AWSTraceHeader – Returns the X-Ray trace header string.

SenderId

For an IAM user, returns the IAM user ID, for example ABCDEFGHI1JKLMNOPQ23R.

For an IAM role, returns the IAM role ID, for example ABCDE1F2GH3I4JK5LMNOP:i-a123b456.

SentTimestamp – Returns the time the message was sent to the queue (epoch time in milliseconds).

MessageDeduplicationId – Returns the value provided by the producer that calls the SendMessage action.

MessageGroupId – Returns the value provided by the producer that calls the SendMessage action. Messages with the same MessageGroupId are returned in sequence.

SequenceNumber – Returns the value provided by Amazon SQS.

receiveMessage_maxNumberOfMessages :: Lens' ReceiveMessage (Maybe Int)

The maximum number of messages to return. Amazon SQS never returns more messages than this value (however, fewer messages might be returned). Valid values: 1 to 10. Default: 1.

receiveMessage_queueUrl :: Lens' ReceiveMessage Text

The URL of the Amazon SQS queue from which messages are received.

Queue URLs and names are case-sensitive.

receiveMessageResponse_messages :: Lens' ReceiveMessageResponse (Maybe [Message])

A list of messages.

receiveMessageResponse_httpStatus :: Lens' ReceiveMessageResponse Int

The response's http status code.

DeleteQueue

deleteQueue_queueUrl :: Lens' DeleteQueue Text

The URL of the Amazon SQS queue to delete.

Queue URLs and names are case-sensitive.

TagQueue

tagQueue_queueUrl :: Lens' TagQueue Text

The URL of the queue.

tagQueue_tags :: Lens' TagQueue (HashMap Text Text)

The list of tags to be added to the specified queue.

DeleteMessageBatch

deleteMessageBatch_queueUrl :: Lens' DeleteMessageBatch Text

The URL of the Amazon SQS queue from which messages are deleted.

Queue URLs and names are case-sensitive.

deleteMessageBatch_entries :: Lens' DeleteMessageBatch [DeleteMessageBatchRequestEntry]

A list of receipt handles for the messages to be deleted.

deleteMessageBatchResponse_httpStatus :: Lens' DeleteMessageBatchResponse Int

The response's http status code.

deleteMessageBatchResponse_successful :: Lens' DeleteMessageBatchResponse [DeleteMessageBatchResultEntry]

A list of DeleteMessageBatchResultEntry items.

deleteMessageBatchResponse_failed :: Lens' DeleteMessageBatchResponse [BatchResultErrorEntry]

A list of BatchResultErrorEntry items.

SetQueueAttributes

setQueueAttributes_queueUrl :: Lens' SetQueueAttributes Text

The URL of the Amazon SQS queue whose attributes are set.

Queue URLs and names are case-sensitive.

setQueueAttributes_attributes :: Lens' SetQueueAttributes (HashMap QueueAttributeName Text)

A map of attributes to set.

The following lists the names, descriptions, and values of the special request parameters that the SetQueueAttributes action uses:

DelaySeconds – The length of time, in seconds, for which the delivery of all messages in the queue is delayed. Valid values: An integer from 0 to 900 (15 minutes). Default: 0.

MaximumMessageSize – The limit of how many bytes a message can contain before Amazon SQS rejects it. Valid values: An integer from 1,024 bytes (1 KiB) up to 262,144 bytes (256 KiB). Default: 262,144 (256 KiB).

MessageRetentionPeriod – The length of time, in seconds, for which Amazon SQS retains a message. Valid values: An integer representing seconds, from 60 (1 minute) to 1,209,600 (14 days). Default: 345,600 (4 days).

Policy – The queue's policy. A valid Amazon Web Services policy. For more information about policy structure, see Overview of Amazon Web Services IAM Policies in the Identity and Access Management User Guide.

ReceiveMessageWaitTimeSeconds – The length of time, in seconds, for which a ReceiveMessage action waits for a message to arrive. Valid values: An integer from 0 to 20 (seconds). Default: 0.

VisibilityTimeout – The visibility timeout for the queue, in seconds. Valid values: An integer from 0 to 43,200 (12 hours). Default: 30. For more information about the visibility timeout, see Visibility Timeout in the Amazon SQS Developer Guide.

The following attributes apply only to dead-letter queues:

RedrivePolicy – The string that includes the parameters for the dead-letter queue functionality of the source queue as a JSON object. The parameters are as follows:

deadLetterTargetArn – The Amazon Resource Name (ARN) of the dead-letter queue to which Amazon SQS moves messages after the value of maxReceiveCount is exceeded.

maxReceiveCount – The number of times a message is delivered to the source queue before being moved to the dead-letter queue. When the ReceiveCount for a message exceeds the maxReceiveCount for a queue, Amazon SQS moves the message to the dead-letter-queue.

RedriveAllowPolicy – The string that includes the parameters for the permissions for the dead-letter queue redrive permission and which source queues can specify dead-letter queues as a JSON object. The parameters are as follows:

redrivePermission – The permission type that defines which source queues can specify the current queue as the dead-letter queue. Valid values are:

allowAll – (Default) Any source queues in this Amazon Web Services account in the same Region can specify this queue as the dead-letter queue.

denyAll – No source queues can specify this queue as the dead-letter queue.

byQueue – Only queues specified by the sourceQueueArns parameter can specify this queue as the dead-letter queue.

sourceQueueArns – The Amazon Resource Names (ARN)s of the source queues that can specify this queue as the dead-letter queue and redrive messages. You can specify this parameter only when the redrivePermission parameter is set to byQueue. You can specify up to 10 source queue ARNs. To allow more than 10 source queues to specify dead-letter queues, set the redrivePermission parameter to allowAll.

The dead-letter queue of a FIFO queue must also be a FIFO queue. Similarly, the dead-letter queue of a standard queue must also be a standard queue.

The following attributes apply only to server-side-encryption:

KmsMasterKeyId – The ID of an Amazon Web Services managed customer master key (CMK) for Amazon SQS or a custom CMK. For more information, see Key Terms. While the alias of the AWS-managed CMK for Amazon SQS is always alias/aws/sqs, the alias of a custom CMK can, for example, be alias/MyAlias . For more examples, see KeyId in the Key Management Service API Reference.

KmsDataKeyReusePeriodSeconds – The length of time, in seconds, for which Amazon SQS can reuse a data key to encrypt or decrypt messages before calling KMS again. An integer representing seconds, between 60 seconds (1 minute) and 86,400 seconds (24 hours). Default: 300 (5 minutes). A shorter time period provides better security but results in more calls to KMS which might incur charges after Free Tier. For more information, see How Does the Data Key Reuse Period Work?.

The following attribute applies only to FIFO (first-in-first-out) queues:

ContentBasedDeduplication – Enables content-based deduplication. For more information, see Exactly-once processing in the Amazon SQS Developer Guide. Note the following:

Every message must have a unique MessageDeduplicationId.

You may provide a MessageDeduplicationId explicitly.

If you aren't able to provide a MessageDeduplicationId and you enable ContentBasedDeduplication for your queue, Amazon SQS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message).

If you don't provide a MessageDeduplicationId and the queue doesn't have ContentBasedDeduplication set, the action fails with an error.

If the queue has ContentBasedDeduplication set, your MessageDeduplicationId overrides the generated one.

When ContentBasedDeduplication is in effect, messages with identical content sent within the deduplication interval are treated as duplicates and only one copy of the message is delivered.

If you send one message with ContentBasedDeduplication enabled and then another message with a MessageDeduplicationId that is the same as the one generated for the first MessageDeduplicationId, the two messages are treated as duplicates and only one copy of the message is delivered.

The following attributes apply only to high throughput for FIFO queues:

DeduplicationScope – Specifies whether message deduplication occurs at the message group or queue level. Valid values are messageGroup and queue.

FifoThroughputLimit – Specifies whether the FIFO queue throughput quota applies to the entire queue or per message group. Valid values are perQueue and perMessageGroupId. The perMessageGroupId value is allowed only when the value for DeduplicationScope is messageGroup.

To enable high throughput for FIFO queues, do the following:

Set DeduplicationScope to messageGroup.

Set FifoThroughputLimit to perMessageGroupId.

If you set these attributes to anything other than the values shown for enabling high throughput, normal throughput is in effect and deduplication occurs as specified.

For information on throughput quotas, see Quotas related to messages in the Amazon SQS Developer Guide.

ListDeadLetterSourceQueues

listDeadLetterSourceQueues_nextToken :: Lens' ListDeadLetterSourceQueues (Maybe Text)

Pagination token to request the next set of results.

listDeadLetterSourceQueues_maxResults :: Lens' ListDeadLetterSourceQueues (Maybe Int)

Maximum number of results to include in the response. Value range is 1 to 1000. You must set MaxResults to receive a value for NextToken in the response.

listDeadLetterSourceQueues_queueUrl :: Lens' ListDeadLetterSourceQueues Text

The URL of a dead-letter queue.

Queue URLs and names are case-sensitive.

listDeadLetterSourceQueuesResponse_nextToken :: Lens' ListDeadLetterSourceQueuesResponse (Maybe Text)

Pagination token to include in the next request. Token value is null if there are no additional results to request, or if you did not set MaxResults in the request.

listDeadLetterSourceQueuesResponse_httpStatus :: Lens' ListDeadLetterSourceQueuesResponse Int

The response's http status code.

listDeadLetterSourceQueuesResponse_queueUrls :: Lens' ListDeadLetterSourceQueuesResponse [Text]

A list of source queue URLs that have the RedrivePolicy queue attribute configured with a dead-letter queue.

AddPermission

addPermission_queueUrl :: Lens' AddPermission Text

The URL of the Amazon SQS queue to which permissions are added.

Queue URLs and names are case-sensitive.

addPermission_label :: Lens' AddPermission Text

The unique identification of the permission you're setting (for example, AliceSendMessage). Maximum 80 characters. Allowed characters include alphanumeric characters, hyphens (-), and underscores (_).

addPermission_aWSAccountIds :: Lens' AddPermission [Text]

The account numbers of the principals who are to receive permission. For information about locating the account identification, see Your Amazon Web Services Identifiers in the Amazon SQS Developer Guide.

addPermission_actions :: Lens' AddPermission [Text]

The action the client wants to allow for the specified principal. Valid values: the name of any action or *.

For more information about these actions, see Overview of Managing Access Permissions to Your Amazon Simple Queue Service Resource in the Amazon SQS Developer Guide.

Specifying SendMessage, DeleteMessage, or ChangeMessageVisibility for ActionName.n also grants permissions for the corresponding batch versions of those actions: SendMessageBatch, DeleteMessageBatch, and ChangeMessageVisibilityBatch.

DeleteMessage

deleteMessage_queueUrl :: Lens' DeleteMessage Text

The URL of the Amazon SQS queue from which messages are deleted.

Queue URLs and names are case-sensitive.

deleteMessage_receiptHandle :: Lens' DeleteMessage Text

The receipt handle associated with the message to delete.

ListQueueTags

listQueueTags_queueUrl :: Lens' ListQueueTags Text

The URL of the queue.

listQueueTagsResponse_tags :: Lens' ListQueueTagsResponse (Maybe (HashMap Text Text))

The list of all tags added to the specified queue.

listQueueTagsResponse_httpStatus :: Lens' ListQueueTagsResponse Int

The response's http status code.

CreateQueue

createQueue_attributes :: Lens' CreateQueue (Maybe (HashMap QueueAttributeName Text))

A map of attributes with their corresponding values.

The following lists the names, descriptions, and values of the special request parameters that the CreateQueue action uses:

DelaySeconds – The length of time, in seconds, for which the delivery of all messages in the queue is delayed. Valid values: An integer from 0 to 900 seconds (15 minutes). Default: 0.

MaximumMessageSize – The limit of how many bytes a message can contain before Amazon SQS rejects it. Valid values: An integer from 1,024 bytes (1 KiB) to 262,144 bytes (256 KiB). Default: 262,144 (256 KiB).

MessageRetentionPeriod – The length of time, in seconds, for which Amazon SQS retains a message. Valid values: An integer from 60 seconds (1 minute) to 1,209,600 seconds (14 days). Default: 345,600 (4 days).

Policy – The queue's policy. A valid Amazon Web Services policy. For more information about policy structure, see Overview of Amazon Web Services IAM Policies in the Amazon IAM User Guide.

ReceiveMessageWaitTimeSeconds – The length of time, in seconds, for which a ReceiveMessage action waits for a message to arrive. Valid values: An integer from 0 to 20 (seconds). Default: 0.

VisibilityTimeout – The visibility timeout for the queue, in seconds. Valid values: An integer from 0 to 43,200 (12 hours). Default: 30. For more information about the visibility timeout, see Visibility Timeout in the Amazon SQS Developer Guide.

The following attributes apply only to dead-letter queues:

RedrivePolicy – The string that includes the parameters for the dead-letter queue functionality of the source queue as a JSON object. The parameters are as follows:

deadLetterTargetArn – The Amazon Resource Name (ARN) of the dead-letter queue to which Amazon SQS moves messages after the value of maxReceiveCount is exceeded.

maxReceiveCount – The number of times a message is delivered to the source queue before being moved to the dead-letter queue. When the ReceiveCount for a message exceeds the maxReceiveCount for a queue, Amazon SQS moves the message to the dead-letter-queue.

RedriveAllowPolicy – The string that includes the parameters for the permissions for the dead-letter queue redrive permission and which source queues can specify dead-letter queues as a JSON object. The parameters are as follows:

redrivePermission – The permission type that defines which source queues can specify the current queue as the dead-letter queue. Valid values are:

allowAll – (Default) Any source queues in this Amazon Web Services account in the same Region can specify this queue as the dead-letter queue.

denyAll – No source queues can specify this queue as the dead-letter queue.

byQueue – Only queues specified by the sourceQueueArns parameter can specify this queue as the dead-letter queue.

sourceQueueArns – The Amazon Resource Names (ARN)s of the source queues that can specify this queue as the dead-letter queue and redrive messages. You can specify this parameter only when the redrivePermission parameter is set to byQueue. You can specify up to 10 source queue ARNs. To allow more than 10 source queues to specify dead-letter queues, set the redrivePermission parameter to allowAll.

The dead-letter queue of a FIFO queue must also be a FIFO queue. Similarly, the dead-letter queue of a standard queue must also be a standard queue.

The following attributes apply only to server-side-encryption:

KmsMasterKeyId – The ID of an Amazon Web Services managed customer master key (CMK) for Amazon SQS or a custom CMK. For more information, see Key Terms. While the alias of the Amazon Web Services managed CMK for Amazon SQS is always alias/aws/sqs, the alias of a custom CMK can, for example, be alias/MyAlias . For more examples, see KeyId in the Key Management Service API Reference.

KmsDataKeyReusePeriodSeconds – The length of time, in seconds, for which Amazon SQS can reuse a data key to encrypt or decrypt messages before calling KMS again. An integer representing seconds, between 60 seconds (1 minute) and 86,400 seconds (24 hours). Default: 300 (5 minutes). A shorter time period provides better security but results in more calls to KMS which might incur charges after Free Tier. For more information, see How Does the Data Key Reuse Period Work?.

The following attributes apply only to FIFO (first-in-first-out) queues:

FifoQueue – Designates a queue as FIFO. Valid values are true and false. If you don't specify the FifoQueue attribute, Amazon SQS creates a standard queue. You can provide this attribute only during queue creation. You can't change it for an existing queue. When you set this attribute, you must also provide the MessageGroupId for your messages explicitly.

For more information, see FIFO queue logic in the Amazon SQS Developer Guide.

ContentBasedDeduplication – Enables content-based deduplication. Valid values are true and false. For more information, see Exactly-once processing in the Amazon SQS Developer Guide. Note the following:

Every message must have a unique MessageDeduplicationId.

You may provide a MessageDeduplicationId explicitly.

If you aren't able to provide a MessageDeduplicationId and you enable ContentBasedDeduplication for your queue, Amazon SQS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message).

If you don't provide a MessageDeduplicationId and the queue doesn't have ContentBasedDeduplication set, the action fails with an error.

If the queue has ContentBasedDeduplication set, your MessageDeduplicationId overrides the generated one.

When ContentBasedDeduplication is in effect, messages with identical content sent within the deduplication interval are treated as duplicates and only one copy of the message is delivered.

If you send one message with ContentBasedDeduplication enabled and then another message with a MessageDeduplicationId that is the same as the one generated for the first MessageDeduplicationId, the two messages are treated as duplicates and only one copy of the message is delivered.

The following attributes apply only to high throughput for FIFO queues:

DeduplicationScope – Specifies whether message deduplication occurs at the message group or queue level. Valid values are messageGroup and queue.

FifoThroughputLimit – Specifies whether the FIFO queue throughput quota applies to the entire queue or per message group. Valid values are perQueue and perMessageGroupId. The perMessageGroupId value is allowed only when the value for DeduplicationScope is messageGroup.

To enable high throughput for FIFO queues, do the following:

Set DeduplicationScope to messageGroup.

Set FifoThroughputLimit to perMessageGroupId.

If you set these attributes to anything other than the values shown for enabling high throughput, normal throughput is in effect and deduplication occurs as specified.

For information on throughput quotas, see Quotas related to messages in the Amazon SQS Developer Guide.

createQueue_tags :: Lens' CreateQueue (Maybe (HashMap Text Text))

Add cost allocation tags to the specified Amazon SQS queue. For an overview, see Tagging Your Amazon SQS Queues in the Amazon SQS Developer Guide.

When you use queue tags, keep the following guidelines in mind:

Adding more than 50 tags to a queue isn't recommended.

Tags don't have any semantic meaning. Amazon SQS interprets tags as character strings.

Tags are case-sensitive.

A new tag with a key identical to that of an existing tag overwrites the existing tag.

For a full list of tag restrictions, see Quotas related to queues in the Amazon SQS Developer Guide.

To be able to tag a queue on creation, you must have the sqs:CreateQueue and sqs:TagQueue permissions.

Cross-account permissions don't apply to this action. For more information, see Grant cross-account permissions to a role and a user name in the Amazon SQS Developer Guide.

createQueue_queueName :: Lens' CreateQueue Text

The name of the new queue. The following limits apply to this name:

A queue name can have up to 80 characters.

Valid values: alphanumeric characters, hyphens (-), and underscores (_).

A FIFO queue name must end with the .fifo suffix.

Queue URLs and names are case-sensitive.

createQueueResponse_queueUrl :: Lens' CreateQueueResponse (Maybe Text)

The URL of the created Amazon SQS queue.

createQueueResponse_httpStatus :: Lens' CreateQueueResponse Int

The response's http status code.

UntagQueue

untagQueue_queueUrl :: Lens' UntagQueue Text

The URL of the queue.

untagQueue_tagKeys :: Lens' UntagQueue [Text]

The list of tags to be removed from the specified queue.

SendMessageBatch

sendMessageBatch_queueUrl :: Lens' SendMessageBatch Text

The URL of the Amazon SQS queue to which batched messages are sent.

Queue URLs and names are case-sensitive.

sendMessageBatch_entries :: Lens' SendMessageBatch [SendMessageBatchRequestEntry]

A list of SendMessageBatchRequestEntry items.

sendMessageBatchResponse_httpStatus :: Lens' SendMessageBatchResponse Int

The response's http status code.

sendMessageBatchResponse_successful :: Lens' SendMessageBatchResponse [SendMessageBatchResultEntry]

A list of SendMessageBatchResultEntry items.

sendMessageBatchResponse_failed :: Lens' SendMessageBatchResponse [BatchResultErrorEntry]

A list of BatchResultErrorEntry items with error details about each message that can't be enqueued.

ChangeMessageVisibility

changeMessageVisibility_queueUrl :: Lens' ChangeMessageVisibility Text

The URL of the Amazon SQS queue whose message's visibility is changed.

Queue URLs and names are case-sensitive.

changeMessageVisibility_receiptHandle :: Lens' ChangeMessageVisibility Text

The receipt handle associated with the message whose visibility timeout is changed. This parameter is returned by the ReceiveMessage action.

changeMessageVisibility_visibilityTimeout :: Lens' ChangeMessageVisibility Int

The new value for the message's visibility timeout (in seconds). Values range: 0 to 43200. Maximum: 12 hours.

Types

BatchResultErrorEntry

batchResultErrorEntry_message :: Lens' BatchResultErrorEntry (Maybe Text)

A message explaining why the action failed on this entry.

batchResultErrorEntry_id :: Lens' BatchResultErrorEntry Text

The Id of an entry in a batch request.

batchResultErrorEntry_senderFault :: Lens' BatchResultErrorEntry Bool

Specifies whether the error happened due to the caller of the batch API action.

batchResultErrorEntry_code :: Lens' BatchResultErrorEntry Text

An error code representing why the action failed on this entry.

ChangeMessageVisibilityBatchRequestEntry

changeMessageVisibilityBatchRequestEntry_visibilityTimeout :: Lens' ChangeMessageVisibilityBatchRequestEntry (Maybe Int)

The new value (in seconds) for the message's visibility timeout.

changeMessageVisibilityBatchRequestEntry_id :: Lens' ChangeMessageVisibilityBatchRequestEntry Text

An identifier for this particular receipt handle used to communicate the result.

The Ids of a batch request need to be unique within a request.

This identifier can have up to 80 characters. The following characters are accepted: alphanumeric characters, hyphens(-), and underscores (_).

changeMessageVisibilityBatchRequestEntry_receiptHandle :: Lens' ChangeMessageVisibilityBatchRequestEntry Text

A receipt handle.

ChangeMessageVisibilityBatchResultEntry

changeMessageVisibilityBatchResultEntry_id :: Lens' ChangeMessageVisibilityBatchResultEntry Text

Represents a message whose visibility timeout has been changed successfully.

DeleteMessageBatchRequestEntry

deleteMessageBatchRequestEntry_id :: Lens' DeleteMessageBatchRequestEntry Text

An identifier for this particular receipt handle. This is used to communicate the result.

The Ids of a batch request need to be unique within a request.

This identifier can have up to 80 characters. The following characters are accepted: alphanumeric characters, hyphens(-), and underscores (_).

deleteMessageBatchRequestEntry_receiptHandle :: Lens' DeleteMessageBatchRequestEntry Text

A receipt handle.

DeleteMessageBatchResultEntry

deleteMessageBatchResultEntry_id :: Lens' DeleteMessageBatchResultEntry Text

Represents a successfully deleted message.

Message

message_messageAttributes :: Lens' Message (Maybe (HashMap Text MessageAttributeValue))

Each message attribute consists of a Name, Type, and Value. For more information, see Amazon SQS message attributes in the Amazon SQS Developer Guide.

message_mD5OfBody :: Lens' Message (Maybe Text)

An MD5 digest of the non-URL-encoded message body string.

message_body :: Lens' Message (Maybe Text)

The message's contents (not URL-encoded).

message_attributes :: Lens' Message (Maybe (HashMap MessageAttribute Text))

A map of the attributes requested in ReceiveMessage to their respective values. Supported attributes:

ApproximateReceiveCount

ApproximateFirstReceiveTimestamp

MessageDeduplicationId

MessageGroupId

SenderId

SentTimestamp

SequenceNumber

ApproximateFirstReceiveTimestamp and SentTimestamp are each returned as an integer representing the epoch time in milliseconds.

message_receiptHandle :: Lens' Message (Maybe Text)

An identifier associated with the act of receiving the message. A new receipt handle is returned every time you receive a message. When deleting a message, you provide the last received receipt handle to delete the message.

message_messageId :: Lens' Message (Maybe Text)

A unique identifier for the message. A MessageIdis considered unique across all accounts for an extended period of time.

message_mD5OfMessageAttributes :: Lens' Message (Maybe Text)

An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

MessageAttributeValue

messageAttributeValue_binaryValue :: Lens' MessageAttributeValue (Maybe ByteString)

Binary type attributes can store any binary data, such as compressed data, encrypted data, or images.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

messageAttributeValue_stringListValues :: Lens' MessageAttributeValue (Maybe [Text])

Not implemented. Reserved for future use.

messageAttributeValue_stringValue :: Lens' MessageAttributeValue (Maybe Text)

Strings are Unicode with UTF-8 binary encoding. For a list of code values, see ASCII Printable Characters.

messageAttributeValue_binaryListValues :: Lens' MessageAttributeValue (Maybe [ByteString])

Not implemented. Reserved for future use.

messageAttributeValue_dataType :: Lens' MessageAttributeValue Text

Amazon SQS supports the following logical data types: String, Number, and Binary. For the Number data type, you must use StringValue.

You can also append custom labels. For more information, see Amazon SQS Message Attributes in the Amazon SQS Developer Guide.

MessageSystemAttributeValue

messageSystemAttributeValue_binaryValue :: Lens' MessageSystemAttributeValue (Maybe ByteString)

Binary type attributes can store any binary data, such as compressed data, encrypted data, or images.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

messageSystemAttributeValue_stringListValues :: Lens' MessageSystemAttributeValue (Maybe [Text])

Not implemented. Reserved for future use.

messageSystemAttributeValue_stringValue :: Lens' MessageSystemAttributeValue (Maybe Text)

Strings are Unicode with UTF-8 binary encoding. For a list of code values, see ASCII Printable Characters.

messageSystemAttributeValue_binaryListValues :: Lens' MessageSystemAttributeValue (Maybe [ByteString])

Not implemented. Reserved for future use.

messageSystemAttributeValue_dataType :: Lens' MessageSystemAttributeValue Text

Amazon SQS supports the following logical data types: String, Number, and Binary. For the Number data type, you must use StringValue.

You can also append custom labels. For more information, see Amazon SQS Message Attributes in the Amazon SQS Developer Guide.

SendMessageBatchRequestEntry

sendMessageBatchRequestEntry_messageAttributes :: Lens' SendMessageBatchRequestEntry (Maybe (HashMap Text MessageAttributeValue))

Each message attribute consists of a Name, Type, and Value. For more information, see Amazon SQS message attributes in the Amazon SQS Developer Guide.

sendMessageBatchRequestEntry_delaySeconds :: Lens' SendMessageBatchRequestEntry (Maybe Int)

The length of time, in seconds, for which a specific message is delayed. Valid values: 0 to 900. Maximum: 15 minutes. Messages with a positive DelaySeconds value become available for processing after the delay period is finished. If you don't specify a value, the default value for the queue is applied.

When you set FifoQueue, you can't set DelaySeconds per message. You can set this parameter only on a queue level.

sendMessageBatchRequestEntry_messageSystemAttributes :: Lens' SendMessageBatchRequestEntry (Maybe (HashMap MessageSystemAttributeNameForSends MessageSystemAttributeValue))

The message system attribute to send Each message system attribute consists of a Name, Type, and Value.

Currently, the only supported message system attribute is AWSTraceHeader. Its type must be String and its value must be a correctly formatted X-Ray trace header string.

The size of a message system attribute doesn't count towards the total size of a message.

sendMessageBatchRequestEntry_messageDeduplicationId :: Lens' SendMessageBatchRequestEntry (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) queues.

The token used for deduplication of messages within a 5-minute minimum deduplication interval. If a message with a particular MessageDeduplicationId is sent successfully, subsequent messages with the same MessageDeduplicationId are accepted successfully but aren't delivered. For more information, see Exactly-once processing in the Amazon SQS Developer Guide.

Every message must have a unique MessageDeduplicationId,

You may provide a MessageDeduplicationId explicitly.

If you aren't able to provide a MessageDeduplicationId and you enable ContentBasedDeduplication for your queue, Amazon SQS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message).

If you don't provide a MessageDeduplicationId and the queue doesn't have ContentBasedDeduplication set, the action fails with an error.

If the queue has ContentBasedDeduplication set, your MessageDeduplicationId overrides the generated one.

When ContentBasedDeduplication is in effect, messages with identical content sent within the deduplication interval are treated as duplicates and only one copy of the message is delivered.

If you send one message with ContentBasedDeduplication enabled and then another message with a MessageDeduplicationId that is the same as the one generated for the first MessageDeduplicationId, the two messages are treated as duplicates and only one copy of the message is delivered.

The MessageDeduplicationId is available to the consumer of the message (this can be useful for troubleshooting delivery issues).

If a message is sent successfully but the acknowledgement is lost and the message is resent with the same MessageDeduplicationId after the deduplication interval, Amazon SQS can't detect duplicate messages.

Amazon SQS continues to keep track of the message deduplication ID even after the message is received and deleted.

The length of MessageDeduplicationId is 128 characters. MessageDeduplicationId can contain alphanumeric characters (a-z, A-Z, 0-9) and punctuation (!"#$%&'()*+,-./:;<=>?@[\]^_`{|}~).

For best practices of using MessageDeduplicationId, see Using the MessageDeduplicationId Property in the Amazon SQS Developer Guide.

sendMessageBatchRequestEntry_messageGroupId :: Lens' SendMessageBatchRequestEntry (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) queues.

The tag that specifies that a message belongs to a specific message group. Messages that belong to the same message group are processed in a FIFO manner (however, messages in different message groups might be processed out of order). To interleave multiple ordered streams within a single queue, use MessageGroupId values (for example, session data for multiple users). In this scenario, multiple consumers can process the queue, but the session data of each user is processed in a FIFO fashion.

You must associate a non-empty MessageGroupId with a message. If you don't provide a MessageGroupId, the action fails.

ReceiveMessage might return messages with multiple MessageGroupId values. For each MessageGroupId, the messages are sorted by time sent. The caller can't specify a MessageGroupId.

The length of MessageGroupId is 128 characters. Valid values: alphanumeric characters and punctuation (!"#$%&'()*+,-./:;<=>?@[\]^_`{|}~).

For best practices of using MessageGroupId, see Using the MessageGroupId Property in the Amazon SQS Developer Guide.

MessageGroupId is required for FIFO queues. You can't use it for Standard queues.

sendMessageBatchRequestEntry_id :: Lens' SendMessageBatchRequestEntry Text

An identifier for a message in this batch used to communicate the result.

The Ids of a batch request need to be unique within a request.

This identifier can have up to 80 characters. The following characters are accepted: alphanumeric characters, hyphens(-), and underscores (_).

sendMessageBatchRequestEntry_messageBody :: Lens' SendMessageBatchRequestEntry Text

The body of the message.

SendMessageBatchResultEntry

sendMessageBatchResultEntry_sequenceNumber :: Lens' SendMessageBatchResultEntry (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) queues.

The large, non-consecutive number that Amazon SQS assigns to each message.

The length of SequenceNumber is 128 bits. As SequenceNumber continues to increase for a particular MessageGroupId.

sendMessageBatchResultEntry_mD5OfMessageSystemAttributes :: Lens' SendMessageBatchResultEntry (Maybe Text)

An MD5 digest of the non-URL-encoded message system attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

sendMessageBatchResultEntry_mD5OfMessageAttributes :: Lens' SendMessageBatchResultEntry (Maybe Text)

An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

sendMessageBatchResultEntry_id :: Lens' SendMessageBatchResultEntry Text

An identifier for the message in this batch.

sendMessageBatchResultEntry_messageId :: Lens' SendMessageBatchResultEntry Text

An identifier for the message.

sendMessageBatchResultEntry_mD5OfMessageBody :: Lens' SendMessageBatchResultEntry Text

An MD5 digest of the non-URL-encoded message body string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.