Network.AWS.SQS.ListQueueTags

Description

List all cost allocation tags added to the specified Amazon SQS queue. For an overview, see Tagging Your Amazon SQS Queues in the Amazon SQS Developer Guide.

Cross-account permissions don't apply to this action. For more information, see Grant cross-account permissions to a role and a user name in the Amazon SQS Developer Guide.

Creating a Request

data ListQueueTags

See: newListQueueTags smart constructor.

Constructors

ListQueueTags'

Fields

queueUrl :: Text

The URL of the queue.

newListQueueTags

Arguments

:: Text
$sel:queueUrl:ListQueueTags'

-> ListQueueTags

Create a value of ListQueueTags with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queueUrl:ListQueueTags', listQueueTags_queueUrl - The URL of the queue.

Request Lenses

listQueueTags_queueUrl :: Lens' ListQueueTags Text

The URL of the queue.

Destructuring the Response

data ListQueueTagsResponse

See: newListQueueTagsResponse smart constructor.

Constructors

ListQueueTagsResponse'

Fields

tags :: Maybe (HashMap Text Text)

The list of all tags added to the specified queue.

httpStatus :: Int

The response's http status code.

newListQueueTagsResponse

Arguments

:: Int
$sel:httpStatus:ListQueueTagsResponse'

-> ListQueueTagsResponse

Create a value of ListQueueTagsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListQueueTagsResponse', listQueueTagsResponse_tags - The list of all tags added to the specified queue.

$sel:httpStatus:ListQueueTagsResponse', listQueueTagsResponse_httpStatus - The response's http status code.

Response Lenses

listQueueTagsResponse_tags :: Lens' ListQueueTagsResponse (Maybe (HashMap Text Text))

The list of all tags added to the specified queue.

listQueueTagsResponse_httpStatus :: Lens' ListQueueTagsResponse Int

The response's http status code.