Network.AWS.SQS.RemovePermission

Description

Revokes any permissions in the queue policy that matches the specified Label parameter.

Only the owner of a queue can remove permissions from it.

Cross-account permissions don't apply to this action. For more information, see Grant cross-account permissions to a role and a user name in the Amazon SQS Developer Guide.

To remove the ability to change queue permissions, you must deny permission to the AddPermission, RemovePermission, and SetQueueAttributes actions in your IAM policy.

Creating a Request

data RemovePermission

See: newRemovePermission smart constructor.

Constructors

RemovePermission'

Fields

queueUrl :: Text

The URL of the Amazon SQS queue from which permissions are removed.

Queue URLs and names are case-sensitive.

label :: Text

The identification of the permission to remove. This is the label added using the AddPermission action.

newRemovePermission

Arguments

:: Text
$sel:queueUrl:RemovePermission'

-> Text
$sel:label:RemovePermission'

-> RemovePermission

Create a value of RemovePermission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queueUrl:RemovePermission', removePermission_queueUrl - The URL of the Amazon SQS queue from which permissions are removed.

Queue URLs and names are case-sensitive.

$sel:label:RemovePermission', removePermission_label - The identification of the permission to remove. This is the label added using the AddPermission action.

Request Lenses

removePermission_queueUrl :: Lens' RemovePermission Text

The URL of the Amazon SQS queue from which permissions are removed.

Queue URLs and names are case-sensitive.

removePermission_label :: Lens' RemovePermission Text

The identification of the permission to remove. This is the label added using the AddPermission action.

Destructuring the Response

data RemovePermissionResponse

See: newRemovePermissionResponse smart constructor.

Constructors

RemovePermissionResponse'

newRemovePermissionResponse :: RemovePermissionResponse

Create a value of RemovePermissionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.