Network.AWS.SQS.TagQueue

Description

Add cost allocation tags to the specified Amazon SQS queue. For an overview, see Tagging Your Amazon SQS Queues in the Amazon SQS Developer Guide.

When you use queue tags, keep the following guidelines in mind:

Adding more than 50 tags to a queue isn't recommended.

Tags don't have any semantic meaning. Amazon SQS interprets tags as character strings.

Tags are case-sensitive.

A new tag with a key identical to that of an existing tag overwrites the existing tag.

For a full list of tag restrictions, see Quotas related to queues in the Amazon SQS Developer Guide.

Cross-account permissions don't apply to this action. For more information, see Grant cross-account permissions to a role and a user name in the Amazon SQS Developer Guide.

Creating a Request

data TagQueue

See: newTagQueue smart constructor.

Constructors

TagQueue'

Fields

queueUrl :: Text

The URL of the queue.

tags :: HashMap Text Text

The list of tags to be added to the specified queue.

newTagQueue

Arguments

:: Text
$sel:queueUrl:TagQueue'

-> TagQueue

Create a value of TagQueue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queueUrl:TagQueue', tagQueue_queueUrl - The URL of the queue.

$sel:tags:TagQueue', tagQueue_tags - The list of tags to be added to the specified queue.

Request Lenses

tagQueue_queueUrl :: Lens' TagQueue Text

The URL of the queue.

tagQueue_tags :: Lens' TagQueue (HashMap Text Text)

The list of tags to be added to the specified queue.

Destructuring the Response

data TagQueueResponse

See: newTagQueueResponse smart constructor.

Constructors

TagQueueResponse'

newTagQueueResponse :: TagQueueResponse

Create a value of TagQueueResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.