Network.AWS.SQS.Types

Description

Service Configuration

defaultService :: Service

API version 2012-11-05 of the Amazon Simple Queue Service SDK configuration.

Errors

_TooManyEntriesInBatchRequest :: AsError a => Getting (First ServiceError) a ServiceError

The batch request contains more entries than permissible.

_InvalidIdFormat :: AsError a => Getting (First ServiceError) a ServiceError

The specified receipt handle isn't valid for the current version.

_QueueNameExists :: AsError a => Getting (First ServiceError) a ServiceError

A queue with this name already exists. Amazon SQS returns this error only if the request includes attributes whose values differ from those of the existing queue.

_EmptyBatchRequest :: AsError a => Getting (First ServiceError) a ServiceError

The batch request doesn't contain any entries.

_UnsupportedOperation :: AsError a => Getting (First ServiceError) a ServiceError

Error code 400. Unsupported operation.

_InvalidMessageContents :: AsError a => Getting (First ServiceError) a ServiceError

The message contains characters outside the allowed set.

_ReceiptHandleIsInvalid :: AsError a => Getting (First ServiceError) a ServiceError

The specified receipt handle isn't valid.

_QueueDeletedRecently :: AsError a => Getting (First ServiceError) a ServiceError

You must wait 60 seconds after deleting a queue before you can create another queue with the same name.

_InvalidAttributeName :: AsError a => Getting (First ServiceError) a ServiceError

The specified attribute doesn't exist.

_InvalidBatchEntryId :: AsError a => Getting (First ServiceError) a ServiceError

The Id of a batch entry in a batch request doesn't abide by the specification.

_BatchEntryIdsNotDistinct :: AsError a => Getting (First ServiceError) a ServiceError

Two or more batch entries in the request have the same Id.

_MessageNotInflight :: AsError a => Getting (First ServiceError) a ServiceError

The specified message isn't in flight.

_PurgeQueueInProgress :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the specified queue previously received a PurgeQueue request within the last 60 seconds (the time it can take to delete the messages in the queue).

_BatchRequestTooLong :: AsError a => Getting (First ServiceError) a ServiceError

The length of all the messages put together is more than the limit.

_OverLimit :: AsError a => Getting (First ServiceError) a ServiceError

The specified action violates a limit. For example, ReceiveMessage returns this error if the maximum number of inflight messages is reached and AddPermission returns this error if the maximum number of permissions for the queue is reached.

_QueueDoesNotExist :: AsError a => Getting (First ServiceError) a ServiceError

The specified queue doesn't exist.

MessageAttribute

newtype MessageAttribute

Constructors

MessageAttribute'

Fields

fromMessageAttribute :: Text

Bundled Patterns

pattern MessageAttribute_All :: MessageAttribute

pattern MessageAttribute_ApproximateFirstReceiveTimestamp :: MessageAttribute

pattern MessageAttribute_ApproximateReceiveCount :: MessageAttribute

pattern MessageAttribute_SenderId :: MessageAttribute

pattern MessageAttribute_SentTimestamp :: MessageAttribute

MessageSystemAttributeNameForSends

newtype MessageSystemAttributeNameForSends

Constructors

MessageSystemAttributeNameForSends'

Fields

fromMessageSystemAttributeNameForSends :: Text

Bundled Patterns

pattern MessageSystemAttributeNameForSends_AWSTraceHeader :: MessageSystemAttributeNameForSends

QueueAttributeName

newtype QueueAttributeName

Constructors

QueueAttributeName'

Fields

fromQueueAttributeName :: Text

Bundled Patterns

pattern QueueAttributeName_All :: QueueAttributeName

pattern QueueAttributeName_ApproximateNumberOfMessages :: QueueAttributeName

pattern QueueAttributeName_ApproximateNumberOfMessagesDelayed :: QueueAttributeName

pattern QueueAttributeName_ApproximateNumberOfMessagesNotVisible :: QueueAttributeName

pattern QueueAttributeName_ContentBasedDeduplication :: QueueAttributeName

pattern QueueAttributeName_CreatedTimestamp :: QueueAttributeName

pattern QueueAttributeName_DeduplicationScope :: QueueAttributeName

pattern QueueAttributeName_DelaySeconds :: QueueAttributeName

pattern QueueAttributeName_FifoQueue :: QueueAttributeName

pattern QueueAttributeName_FifoThroughputLimit :: QueueAttributeName

pattern QueueAttributeName_KmsDataKeyReusePeriodSeconds :: QueueAttributeName

pattern QueueAttributeName_KmsMasterKeyId :: QueueAttributeName

pattern QueueAttributeName_LastModifiedTimestamp :: QueueAttributeName

pattern QueueAttributeName_MaximumMessageSize :: QueueAttributeName

pattern QueueAttributeName_MessageRetentionPeriod :: QueueAttributeName

pattern QueueAttributeName_Policy :: QueueAttributeName

pattern QueueAttributeName_QueueArn :: QueueAttributeName

pattern QueueAttributeName_ReceiveMessageWaitTimeSeconds :: QueueAttributeName

pattern QueueAttributeName_RedriveAllowPolicy :: QueueAttributeName

pattern QueueAttributeName_RedrivePolicy :: QueueAttributeName

pattern QueueAttributeName_VisibilityTimeout :: QueueAttributeName

BatchResultErrorEntry

data BatchResultErrorEntry

Gives a detailed description of the result of an action on each entry in the request.

See: newBatchResultErrorEntry smart constructor.

Constructors

BatchResultErrorEntry'

Fields

message :: Maybe Text

A message explaining why the action failed on this entry.

id :: Text

The Id of an entry in a batch request.

senderFault :: Bool

Specifies whether the error happened due to the caller of the batch API action.

code :: Text

An error code representing why the action failed on this entry.

newBatchResultErrorEntry

Arguments

:: Text
$sel:id:BatchResultErrorEntry'

-> Bool
$sel:senderFault:BatchResultErrorEntry'

-> Text
$sel:code:BatchResultErrorEntry'

-> BatchResultErrorEntry

Create a value of BatchResultErrorEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:BatchResultErrorEntry', batchResultErrorEntry_message - A message explaining why the action failed on this entry.

$sel:id:BatchResultErrorEntry', batchResultErrorEntry_id - The Id of an entry in a batch request.

$sel:senderFault:BatchResultErrorEntry', batchResultErrorEntry_senderFault - Specifies whether the error happened due to the caller of the batch API action.

$sel:code:BatchResultErrorEntry', batchResultErrorEntry_code - An error code representing why the action failed on this entry.

batchResultErrorEntry_message :: Lens' BatchResultErrorEntry (Maybe Text)

A message explaining why the action failed on this entry.

batchResultErrorEntry_id :: Lens' BatchResultErrorEntry Text

The Id of an entry in a batch request.

batchResultErrorEntry_senderFault :: Lens' BatchResultErrorEntry Bool

Specifies whether the error happened due to the caller of the batch API action.

batchResultErrorEntry_code :: Lens' BatchResultErrorEntry Text

An error code representing why the action failed on this entry.

ChangeMessageVisibilityBatchRequestEntry

data ChangeMessageVisibilityBatchRequestEntry

Encloses a receipt handle and an entry id for each message in ChangeMessageVisibilityBatch.

All of the following list parameters must be prefixed with ChangeMessageVisibilityBatchRequestEntry.n, where n is an integer value starting with 1. For example, a parameter list for this action might look like this:

&ChangeMessageVisibilityBatchRequestEntry.1.Id=change_visibility_msg_2

&ChangeMessageVisibilityBatchRequestEntry.1.ReceiptHandle=your_receipt_handle

&ChangeMessageVisibilityBatchRequestEntry.1.VisibilityTimeout=45

See: newChangeMessageVisibilityBatchRequestEntry smart constructor.

Constructors

ChangeMessageVisibilityBatchRequestEntry'

Fields

visibilityTimeout :: Maybe Int

The new value (in seconds) for the message's visibility timeout.

id :: Text

An identifier for this particular receipt handle used to communicate the result.

The Ids of a batch request need to be unique within a request.

This identifier can have up to 80 characters. The following characters are accepted: alphanumeric characters, hyphens(-), and underscores (_).

receiptHandle :: Text

A receipt handle.

newChangeMessageVisibilityBatchRequestEntry

Arguments

:: Text
$sel:id:ChangeMessageVisibilityBatchRequestEntry'

-> Text
$sel:receiptHandle:ChangeMessageVisibilityBatchRequestEntry'

-> ChangeMessageVisibilityBatchRequestEntry

Create a value of ChangeMessageVisibilityBatchRequestEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:visibilityTimeout:ChangeMessageVisibilityBatchRequestEntry', changeMessageVisibilityBatchRequestEntry_visibilityTimeout - The new value (in seconds) for the message's visibility timeout.

$sel:id:ChangeMessageVisibilityBatchRequestEntry', changeMessageVisibilityBatchRequestEntry_id - An identifier for this particular receipt handle used to communicate the result.

The Ids of a batch request need to be unique within a request.

This identifier can have up to 80 characters. The following characters are accepted: alphanumeric characters, hyphens(-), and underscores (_).

$sel:receiptHandle:ChangeMessageVisibilityBatchRequestEntry', changeMessageVisibilityBatchRequestEntry_receiptHandle - A receipt handle.

changeMessageVisibilityBatchRequestEntry_visibilityTimeout :: Lens' ChangeMessageVisibilityBatchRequestEntry (Maybe Int)

The new value (in seconds) for the message's visibility timeout.

changeMessageVisibilityBatchRequestEntry_id :: Lens' ChangeMessageVisibilityBatchRequestEntry Text

An identifier for this particular receipt handle used to communicate the result.

The Ids of a batch request need to be unique within a request.

This identifier can have up to 80 characters. The following characters are accepted: alphanumeric characters, hyphens(-), and underscores (_).

changeMessageVisibilityBatchRequestEntry_receiptHandle :: Lens' ChangeMessageVisibilityBatchRequestEntry Text

A receipt handle.

ChangeMessageVisibilityBatchResultEntry

data ChangeMessageVisibilityBatchResultEntry

Encloses the Id of an entry in ChangeMessageVisibilityBatch.

See: newChangeMessageVisibilityBatchResultEntry smart constructor.

Constructors

ChangeMessageVisibilityBatchResultEntry'

Fields

id :: Text

Represents a message whose visibility timeout has been changed successfully.

newChangeMessageVisibilityBatchResultEntry

Arguments

:: Text
$sel:id:ChangeMessageVisibilityBatchResultEntry'

-> ChangeMessageVisibilityBatchResultEntry

Create a value of ChangeMessageVisibilityBatchResultEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:ChangeMessageVisibilityBatchResultEntry', changeMessageVisibilityBatchResultEntry_id - Represents a message whose visibility timeout has been changed successfully.

changeMessageVisibilityBatchResultEntry_id :: Lens' ChangeMessageVisibilityBatchResultEntry Text

Represents a message whose visibility timeout has been changed successfully.

DeleteMessageBatchRequestEntry

data DeleteMessageBatchRequestEntry

Encloses a receipt handle and an identifier for it.

See: newDeleteMessageBatchRequestEntry smart constructor.

Constructors

DeleteMessageBatchRequestEntry'

Fields

id :: Text

An identifier for this particular receipt handle. This is used to communicate the result.

The Ids of a batch request need to be unique within a request.

This identifier can have up to 80 characters. The following characters are accepted: alphanumeric characters, hyphens(-), and underscores (_).

receiptHandle :: Text

A receipt handle.

newDeleteMessageBatchRequestEntry

Arguments

:: Text
$sel:id:DeleteMessageBatchRequestEntry'

-> Text
$sel:receiptHandle:DeleteMessageBatchRequestEntry'

-> DeleteMessageBatchRequestEntry

Create a value of DeleteMessageBatchRequestEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DeleteMessageBatchRequestEntry', deleteMessageBatchRequestEntry_id - An identifier for this particular receipt handle. This is used to communicate the result.

The Ids of a batch request need to be unique within a request.

This identifier can have up to 80 characters. The following characters are accepted: alphanumeric characters, hyphens(-), and underscores (_).

$sel:receiptHandle:DeleteMessageBatchRequestEntry', deleteMessageBatchRequestEntry_receiptHandle - A receipt handle.

deleteMessageBatchRequestEntry_id :: Lens' DeleteMessageBatchRequestEntry Text

An identifier for this particular receipt handle. This is used to communicate the result.

The Ids of a batch request need to be unique within a request.

This identifier can have up to 80 characters. The following characters are accepted: alphanumeric characters, hyphens(-), and underscores (_).

deleteMessageBatchRequestEntry_receiptHandle :: Lens' DeleteMessageBatchRequestEntry Text

A receipt handle.

DeleteMessageBatchResultEntry

data DeleteMessageBatchResultEntry

Encloses the Id of an entry in DeleteMessageBatch.

See: newDeleteMessageBatchResultEntry smart constructor.

Constructors

DeleteMessageBatchResultEntry'

Fields

id :: Text

Represents a successfully deleted message.

newDeleteMessageBatchResultEntry

Arguments

:: Text
$sel:id:DeleteMessageBatchResultEntry'

-> DeleteMessageBatchResultEntry

Create a value of DeleteMessageBatchResultEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DeleteMessageBatchResultEntry', deleteMessageBatchResultEntry_id - Represents a successfully deleted message.

deleteMessageBatchResultEntry_id :: Lens' DeleteMessageBatchResultEntry Text

Represents a successfully deleted message.

Message

data Message

An Amazon SQS message.

See: newMessage smart constructor.

Constructors

Message'

Fields

body :: Maybe Text

The message's contents (not URL-encoded).

mD5OfBody :: Maybe Text

An MD5 digest of the non-URL-encoded message body string.

attributes :: Maybe (HashMap MessageAttribute Text)

A map of the attributes requested in ReceiveMessage to their respective values. Supported attributes:

ApproximateReceiveCount

ApproximateFirstReceiveTimestamp

MessageDeduplicationId

MessageGroupId

SenderId

SentTimestamp

SequenceNumber

ApproximateFirstReceiveTimestamp and SentTimestamp are each returned as an integer representing the epoch time in milliseconds.

messageAttributes :: Maybe (HashMap Text MessageAttributeValue)

Each message attribute consists of a Name, Type, and Value. For more information, see Amazon SQS message attributes in the Amazon SQS Developer Guide.

mD5OfMessageAttributes :: Maybe Text

An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

receiptHandle :: Maybe Text

An identifier associated with the act of receiving the message. A new receipt handle is returned every time you receive a message. When deleting a message, you provide the last received receipt handle to delete the message.

messageId :: Maybe Text

A unique identifier for the message. A MessageIdis considered unique across all accounts for an extended period of time.

newMessage :: Message

Create a value of Message with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:body:Message', message_body - The message's contents (not URL-encoded).

$sel:mD5OfBody:Message', message_mD5OfBody - An MD5 digest of the non-URL-encoded message body string.

$sel:attributes:Message', message_attributes - A map of the attributes requested in ReceiveMessage to their respective values. Supported attributes:

ApproximateReceiveCount

ApproximateFirstReceiveTimestamp

MessageDeduplicationId

MessageGroupId

SenderId

SentTimestamp

SequenceNumber

ApproximateFirstReceiveTimestamp and SentTimestamp are each returned as an integer representing the epoch time in milliseconds.

$sel:messageAttributes:Message', message_messageAttributes - Each message attribute consists of a Name, Type, and Value. For more information, see Amazon SQS message attributes in the Amazon SQS Developer Guide.

$sel:mD5OfMessageAttributes:Message', message_mD5OfMessageAttributes - An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

$sel:receiptHandle:Message', message_receiptHandle - An identifier associated with the act of receiving the message. A new receipt handle is returned every time you receive a message. When deleting a message, you provide the last received receipt handle to delete the message.

$sel:messageId:Message', message_messageId - A unique identifier for the message. A MessageIdis considered unique across all accounts for an extended period of time.

message_body :: Lens' Message (Maybe Text)

The message's contents (not URL-encoded).

message_mD5OfBody :: Lens' Message (Maybe Text)

An MD5 digest of the non-URL-encoded message body string.

message_attributes :: Lens' Message (Maybe (HashMap MessageAttribute Text))

A map of the attributes requested in ReceiveMessage to their respective values. Supported attributes:

ApproximateReceiveCount

ApproximateFirstReceiveTimestamp

MessageDeduplicationId

MessageGroupId

SenderId

SentTimestamp

SequenceNumber

ApproximateFirstReceiveTimestamp and SentTimestamp are each returned as an integer representing the epoch time in milliseconds.

message_messageAttributes :: Lens' Message (Maybe (HashMap Text MessageAttributeValue))

Each message attribute consists of a Name, Type, and Value. For more information, see Amazon SQS message attributes in the Amazon SQS Developer Guide.

message_mD5OfMessageAttributes :: Lens' Message (Maybe Text)

An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

message_receiptHandle :: Lens' Message (Maybe Text)

An identifier associated with the act of receiving the message. A new receipt handle is returned every time you receive a message. When deleting a message, you provide the last received receipt handle to delete the message.

message_messageId :: Lens' Message (Maybe Text)

A unique identifier for the message. A MessageIdis considered unique across all accounts for an extended period of time.

MessageAttributeValue

data MessageAttributeValue

The user-specified message attribute value. For string data types, the Value attribute has the same restrictions on the content as the message body. For more information, see SendMessage.

Name, type, value and the message body must not be empty or null. All parts of the message attribute, including Name, Type, and Value, are part of the message size restriction (256 KB or 262,144 bytes).

See: newMessageAttributeValue smart constructor.

Constructors

MessageAttributeValue'

Fields

stringListValues :: Maybe [Text]

Not implemented. Reserved for future use.

stringValue :: Maybe Text

Strings are Unicode with UTF-8 binary encoding. For a list of code values, see ASCII Printable Characters.

binaryListValues :: Maybe [Base64]

Not implemented. Reserved for future use.

binaryValue :: Maybe Base64

Binary type attributes can store any binary data, such as compressed data, encrypted data, or images.

dataType :: Text

Amazon SQS supports the following logical data types: String, Number, and Binary. For the Number data type, you must use StringValue.

You can also append custom labels. For more information, see Amazon SQS Message Attributes in the Amazon SQS Developer Guide.

newMessageAttributeValue

Arguments

:: Text
$sel:dataType:MessageAttributeValue'

-> MessageAttributeValue

Create a value of MessageAttributeValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stringListValues:MessageAttributeValue', messageAttributeValue_stringListValues - Not implemented. Reserved for future use.

$sel:stringValue:MessageAttributeValue', messageAttributeValue_stringValue - Strings are Unicode with UTF-8 binary encoding. For a list of code values, see ASCII Printable Characters.

$sel:binaryListValues:MessageAttributeValue', messageAttributeValue_binaryListValues - Not implemented. Reserved for future use.

$sel:binaryValue:MessageAttributeValue', messageAttributeValue_binaryValue - Binary type attributes can store any binary data, such as compressed data, encrypted data, or images.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:dataType:MessageAttributeValue', messageAttributeValue_dataType - Amazon SQS supports the following logical data types: String, Number, and Binary. For the Number data type, you must use StringValue.

You can also append custom labels. For more information, see Amazon SQS Message Attributes in the Amazon SQS Developer Guide.

messageAttributeValue_stringListValues :: Lens' MessageAttributeValue (Maybe [Text])

Not implemented. Reserved for future use.

messageAttributeValue_stringValue :: Lens' MessageAttributeValue (Maybe Text)

Strings are Unicode with UTF-8 binary encoding. For a list of code values, see ASCII Printable Characters.

messageAttributeValue_binaryListValues :: Lens' MessageAttributeValue (Maybe [ByteString])

Not implemented. Reserved for future use.

messageAttributeValue_binaryValue :: Lens' MessageAttributeValue (Maybe ByteString)

Binary type attributes can store any binary data, such as compressed data, encrypted data, or images.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

messageAttributeValue_dataType :: Lens' MessageAttributeValue Text

Amazon SQS supports the following logical data types: String, Number, and Binary. For the Number data type, you must use StringValue.

You can also append custom labels. For more information, see Amazon SQS Message Attributes in the Amazon SQS Developer Guide.

MessageSystemAttributeValue

data MessageSystemAttributeValue

The user-specified message system attribute value. For string data types, the Value attribute has the same restrictions on the content as the message body. For more information, see SendMessage.

Name, type, value and the message body must not be empty or null.

See: newMessageSystemAttributeValue smart constructor.

Constructors

MessageSystemAttributeValue'

Fields

stringListValues :: Maybe [Text]

Not implemented. Reserved for future use.

stringValue :: Maybe Text

Strings are Unicode with UTF-8 binary encoding. For a list of code values, see ASCII Printable Characters.

binaryListValues :: Maybe [Base64]

Not implemented. Reserved for future use.

binaryValue :: Maybe Base64

Binary type attributes can store any binary data, such as compressed data, encrypted data, or images.

dataType :: Text

Amazon SQS supports the following logical data types: String, Number, and Binary. For the Number data type, you must use StringValue.

You can also append custom labels. For more information, see Amazon SQS Message Attributes in the Amazon SQS Developer Guide.

newMessageSystemAttributeValue

Arguments

:: Text
$sel:dataType:MessageSystemAttributeValue'

-> MessageSystemAttributeValue

Create a value of MessageSystemAttributeValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stringListValues:MessageSystemAttributeValue', messageSystemAttributeValue_stringListValues - Not implemented. Reserved for future use.

$sel:stringValue:MessageSystemAttributeValue', messageSystemAttributeValue_stringValue - Strings are Unicode with UTF-8 binary encoding. For a list of code values, see ASCII Printable Characters.

$sel:binaryListValues:MessageSystemAttributeValue', messageSystemAttributeValue_binaryListValues - Not implemented. Reserved for future use.

$sel:binaryValue:MessageSystemAttributeValue', messageSystemAttributeValue_binaryValue - Binary type attributes can store any binary data, such as compressed data, encrypted data, or images.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:dataType:MessageSystemAttributeValue', messageSystemAttributeValue_dataType - Amazon SQS supports the following logical data types: String, Number, and Binary. For the Number data type, you must use StringValue.

You can also append custom labels. For more information, see Amazon SQS Message Attributes in the Amazon SQS Developer Guide.

messageSystemAttributeValue_stringListValues :: Lens' MessageSystemAttributeValue (Maybe [Text])

Not implemented. Reserved for future use.

messageSystemAttributeValue_stringValue :: Lens' MessageSystemAttributeValue (Maybe Text)

Strings are Unicode with UTF-8 binary encoding. For a list of code values, see ASCII Printable Characters.

messageSystemAttributeValue_binaryListValues :: Lens' MessageSystemAttributeValue (Maybe [ByteString])

Not implemented. Reserved for future use.

messageSystemAttributeValue_binaryValue :: Lens' MessageSystemAttributeValue (Maybe ByteString)

Binary type attributes can store any binary data, such as compressed data, encrypted data, or images.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

messageSystemAttributeValue_dataType :: Lens' MessageSystemAttributeValue Text

Amazon SQS supports the following logical data types: String, Number, and Binary. For the Number data type, you must use StringValue.

You can also append custom labels. For more information, see Amazon SQS Message Attributes in the Amazon SQS Developer Guide.

SendMessageBatchRequestEntry

data SendMessageBatchRequestEntry

Contains the details of a single Amazon SQS message along with an Id.

See: newSendMessageBatchRequestEntry smart constructor.

Constructors

SendMessageBatchRequestEntry'

Fields

messageDeduplicationId :: Maybe Text

This parameter applies only to FIFO (first-in-first-out) queues.

The token used for deduplication of messages within a 5-minute minimum deduplication interval. If a message with a particular MessageDeduplicationId is sent successfully, subsequent messages with the same MessageDeduplicationId are accepted successfully but aren't delivered. For more information, see Exactly-once processing in the Amazon SQS Developer Guide.

Every message must have a unique MessageDeduplicationId,

You may provide a MessageDeduplicationId explicitly.

If you aren't able to provide a MessageDeduplicationId and you enable ContentBasedDeduplication for your queue, Amazon SQS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message).

If you don't provide a MessageDeduplicationId and the queue doesn't have ContentBasedDeduplication set, the action fails with an error.

If the queue has ContentBasedDeduplication set, your MessageDeduplicationId overrides the generated one.

When ContentBasedDeduplication is in effect, messages with identical content sent within the deduplication interval are treated as duplicates and only one copy of the message is delivered.

If you send one message with ContentBasedDeduplication enabled and then another message with a MessageDeduplicationId that is the same as the one generated for the first MessageDeduplicationId, the two messages are treated as duplicates and only one copy of the message is delivered.

The MessageDeduplicationId is available to the consumer of the message (this can be useful for troubleshooting delivery issues).

If a message is sent successfully but the acknowledgement is lost and the message is resent with the same MessageDeduplicationId after the deduplication interval, Amazon SQS can't detect duplicate messages.

Amazon SQS continues to keep track of the message deduplication ID even after the message is received and deleted.

The length of MessageDeduplicationId is 128 characters. MessageDeduplicationId can contain alphanumeric characters (a-z, A-Z, 0-9) and punctuation (!"#$%&'()*+,-./:;<=>?@[\]^_`{|}~).

For best practices of using MessageDeduplicationId, see Using the MessageDeduplicationId Property in the Amazon SQS Developer Guide.

messageAttributes :: Maybe (HashMap Text MessageAttributeValue)

Each message attribute consists of a Name, Type, and Value. For more information, see Amazon SQS message attributes in the Amazon SQS Developer Guide.

messageGroupId :: Maybe Text

This parameter applies only to FIFO (first-in-first-out) queues.

The tag that specifies that a message belongs to a specific message group. Messages that belong to the same message group are processed in a FIFO manner (however, messages in different message groups might be processed out of order). To interleave multiple ordered streams within a single queue, use MessageGroupId values (for example, session data for multiple users). In this scenario, multiple consumers can process the queue, but the session data of each user is processed in a FIFO fashion.

You must associate a non-empty MessageGroupId with a message. If you don't provide a MessageGroupId, the action fails.

ReceiveMessage might return messages with multiple MessageGroupId values. For each MessageGroupId, the messages are sorted by time sent. The caller can't specify a MessageGroupId.

The length of MessageGroupId is 128 characters. Valid values: alphanumeric characters and punctuation (!"#$%&'()*+,-./:;<=>?@[\]^_`{|}~).

For best practices of using MessageGroupId, see Using the MessageGroupId Property in the Amazon SQS Developer Guide.

MessageGroupId is required for FIFO queues. You can't use it for Standard queues.

messageSystemAttributes :: Maybe (HashMap MessageSystemAttributeNameForSends MessageSystemAttributeValue)

The message system attribute to send Each message system attribute consists of a Name, Type, and Value.

Currently, the only supported message system attribute is AWSTraceHeader. Its type must be String and its value must be a correctly formatted X-Ray trace header string.

The size of a message system attribute doesn't count towards the total size of a message.

delaySeconds :: Maybe Int

The length of time, in seconds, for which a specific message is delayed. Valid values: 0 to 900. Maximum: 15 minutes. Messages with a positive DelaySeconds value become available for processing after the delay period is finished. If you don't specify a value, the default value for the queue is applied.

When you set FifoQueue, you can't set DelaySeconds per message. You can set this parameter only on a queue level.

id :: Text

An identifier for a message in this batch used to communicate the result.

The Ids of a batch request need to be unique within a request.

This identifier can have up to 80 characters. The following characters are accepted: alphanumeric characters, hyphens(-), and underscores (_).

messageBody :: Text

The body of the message.

newSendMessageBatchRequestEntry

Arguments

:: Text
$sel:id:SendMessageBatchRequestEntry'

-> Text
$sel:messageBody:SendMessageBatchRequestEntry'

-> SendMessageBatchRequestEntry

Create a value of SendMessageBatchRequestEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:messageDeduplicationId:SendMessageBatchRequestEntry', sendMessageBatchRequestEntry_messageDeduplicationId - This parameter applies only to FIFO (first-in-first-out) queues.

The token used for deduplication of messages within a 5-minute minimum deduplication interval. If a message with a particular MessageDeduplicationId is sent successfully, subsequent messages with the same MessageDeduplicationId are accepted successfully but aren't delivered. For more information, see Exactly-once processing in the Amazon SQS Developer Guide.

Every message must have a unique MessageDeduplicationId,

You may provide a MessageDeduplicationId explicitly.

If you aren't able to provide a MessageDeduplicationId and you enable ContentBasedDeduplication for your queue, Amazon SQS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message).

If you don't provide a MessageDeduplicationId and the queue doesn't have ContentBasedDeduplication set, the action fails with an error.

If the queue has ContentBasedDeduplication set, your MessageDeduplicationId overrides the generated one.

When ContentBasedDeduplication is in effect, messages with identical content sent within the deduplication interval are treated as duplicates and only one copy of the message is delivered.

If you send one message with ContentBasedDeduplication enabled and then another message with a MessageDeduplicationId that is the same as the one generated for the first MessageDeduplicationId, the two messages are treated as duplicates and only one copy of the message is delivered.

The MessageDeduplicationId is available to the consumer of the message (this can be useful for troubleshooting delivery issues).

If a message is sent successfully but the acknowledgement is lost and the message is resent with the same MessageDeduplicationId after the deduplication interval, Amazon SQS can't detect duplicate messages.

Amazon SQS continues to keep track of the message deduplication ID even after the message is received and deleted.

The length of MessageDeduplicationId is 128 characters. MessageDeduplicationId can contain alphanumeric characters (a-z, A-Z, 0-9) and punctuation (!"#$%&'()*+,-./:;<=>?@[\]^_`{|}~).

For best practices of using MessageDeduplicationId, see Using the MessageDeduplicationId Property in the Amazon SQS Developer Guide.

$sel:messageAttributes:SendMessageBatchRequestEntry', sendMessageBatchRequestEntry_messageAttributes - Each message attribute consists of a Name, Type, and Value. For more information, see Amazon SQS message attributes in the Amazon SQS Developer Guide.

$sel:messageGroupId:SendMessageBatchRequestEntry', sendMessageBatchRequestEntry_messageGroupId - This parameter applies only to FIFO (first-in-first-out) queues.

The tag that specifies that a message belongs to a specific message group. Messages that belong to the same message group are processed in a FIFO manner (however, messages in different message groups might be processed out of order). To interleave multiple ordered streams within a single queue, use MessageGroupId values (for example, session data for multiple users). In this scenario, multiple consumers can process the queue, but the session data of each user is processed in a FIFO fashion.

You must associate a non-empty MessageGroupId with a message. If you don't provide a MessageGroupId, the action fails.

ReceiveMessage might return messages with multiple MessageGroupId values. For each MessageGroupId, the messages are sorted by time sent. The caller can't specify a MessageGroupId.

The length of MessageGroupId is 128 characters. Valid values: alphanumeric characters and punctuation (!"#$%&'()*+,-./:;<=>?@[\]^_`{|}~).

For best practices of using MessageGroupId, see Using the MessageGroupId Property in the Amazon SQS Developer Guide.

MessageGroupId is required for FIFO queues. You can't use it for Standard queues.

$sel:messageSystemAttributes:SendMessageBatchRequestEntry', sendMessageBatchRequestEntry_messageSystemAttributes - The message system attribute to send Each message system attribute consists of a Name, Type, and Value.

Currently, the only supported message system attribute is AWSTraceHeader. Its type must be String and its value must be a correctly formatted X-Ray trace header string.

The size of a message system attribute doesn't count towards the total size of a message.

$sel:delaySeconds:SendMessageBatchRequestEntry', sendMessageBatchRequestEntry_delaySeconds - The length of time, in seconds, for which a specific message is delayed. Valid values: 0 to 900. Maximum: 15 minutes. Messages with a positive DelaySeconds value become available for processing after the delay period is finished. If you don't specify a value, the default value for the queue is applied.

When you set FifoQueue, you can't set DelaySeconds per message. You can set this parameter only on a queue level.

$sel:id:SendMessageBatchRequestEntry', sendMessageBatchRequestEntry_id - An identifier for a message in this batch used to communicate the result.

The Ids of a batch request need to be unique within a request.

This identifier can have up to 80 characters. The following characters are accepted: alphanumeric characters, hyphens(-), and underscores (_).

$sel:messageBody:SendMessageBatchRequestEntry', sendMessageBatchRequestEntry_messageBody - The body of the message.

sendMessageBatchRequestEntry_messageDeduplicationId :: Lens' SendMessageBatchRequestEntry (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) queues.

The token used for deduplication of messages within a 5-minute minimum deduplication interval. If a message with a particular MessageDeduplicationId is sent successfully, subsequent messages with the same MessageDeduplicationId are accepted successfully but aren't delivered. For more information, see Exactly-once processing in the Amazon SQS Developer Guide.

Every message must have a unique MessageDeduplicationId,

You may provide a MessageDeduplicationId explicitly.

If you aren't able to provide a MessageDeduplicationId and you enable ContentBasedDeduplication for your queue, Amazon SQS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message).

If you don't provide a MessageDeduplicationId and the queue doesn't have ContentBasedDeduplication set, the action fails with an error.

If the queue has ContentBasedDeduplication set, your MessageDeduplicationId overrides the generated one.

When ContentBasedDeduplication is in effect, messages with identical content sent within the deduplication interval are treated as duplicates and only one copy of the message is delivered.

If you send one message with ContentBasedDeduplication enabled and then another message with a MessageDeduplicationId that is the same as the one generated for the first MessageDeduplicationId, the two messages are treated as duplicates and only one copy of the message is delivered.

The MessageDeduplicationId is available to the consumer of the message (this can be useful for troubleshooting delivery issues).

If a message is sent successfully but the acknowledgement is lost and the message is resent with the same MessageDeduplicationId after the deduplication interval, Amazon SQS can't detect duplicate messages.

Amazon SQS continues to keep track of the message deduplication ID even after the message is received and deleted.

The length of MessageDeduplicationId is 128 characters. MessageDeduplicationId can contain alphanumeric characters (a-z, A-Z, 0-9) and punctuation (!"#$%&'()*+,-./:;<=>?@[\]^_`{|}~).

For best practices of using MessageDeduplicationId, see Using the MessageDeduplicationId Property in the Amazon SQS Developer Guide.

sendMessageBatchRequestEntry_messageAttributes :: Lens' SendMessageBatchRequestEntry (Maybe (HashMap Text MessageAttributeValue))

Each message attribute consists of a Name, Type, and Value. For more information, see Amazon SQS message attributes in the Amazon SQS Developer Guide.

sendMessageBatchRequestEntry_messageGroupId :: Lens' SendMessageBatchRequestEntry (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) queues.

The tag that specifies that a message belongs to a specific message group. Messages that belong to the same message group are processed in a FIFO manner (however, messages in different message groups might be processed out of order). To interleave multiple ordered streams within a single queue, use MessageGroupId values (for example, session data for multiple users). In this scenario, multiple consumers can process the queue, but the session data of each user is processed in a FIFO fashion.

You must associate a non-empty MessageGroupId with a message. If you don't provide a MessageGroupId, the action fails.

ReceiveMessage might return messages with multiple MessageGroupId values. For each MessageGroupId, the messages are sorted by time sent. The caller can't specify a MessageGroupId.

The length of MessageGroupId is 128 characters. Valid values: alphanumeric characters and punctuation (!"#$%&'()*+,-./:;<=>?@[\]^_`{|}~).

For best practices of using MessageGroupId, see Using the MessageGroupId Property in the Amazon SQS Developer Guide.

MessageGroupId is required for FIFO queues. You can't use it for Standard queues.

sendMessageBatchRequestEntry_messageSystemAttributes :: Lens' SendMessageBatchRequestEntry (Maybe (HashMap MessageSystemAttributeNameForSends MessageSystemAttributeValue))

The message system attribute to send Each message system attribute consists of a Name, Type, and Value.

Currently, the only supported message system attribute is AWSTraceHeader. Its type must be String and its value must be a correctly formatted X-Ray trace header string.

The size of a message system attribute doesn't count towards the total size of a message.

sendMessageBatchRequestEntry_delaySeconds :: Lens' SendMessageBatchRequestEntry (Maybe Int)

The length of time, in seconds, for which a specific message is delayed. Valid values: 0 to 900. Maximum: 15 minutes. Messages with a positive DelaySeconds value become available for processing after the delay period is finished. If you don't specify a value, the default value for the queue is applied.

When you set FifoQueue, you can't set DelaySeconds per message. You can set this parameter only on a queue level.

sendMessageBatchRequestEntry_id :: Lens' SendMessageBatchRequestEntry Text

An identifier for a message in this batch used to communicate the result.

The Ids of a batch request need to be unique within a request.

This identifier can have up to 80 characters. The following characters are accepted: alphanumeric characters, hyphens(-), and underscores (_).

sendMessageBatchRequestEntry_messageBody :: Lens' SendMessageBatchRequestEntry Text

The body of the message.

SendMessageBatchResultEntry

data SendMessageBatchResultEntry

Encloses a MessageId for a successfully-enqueued message in a SendMessageBatch.

See: newSendMessageBatchResultEntry smart constructor.

Constructors

SendMessageBatchResultEntry'

Fields

sequenceNumber :: Maybe Text

This parameter applies only to FIFO (first-in-first-out) queues.

The large, non-consecutive number that Amazon SQS assigns to each message.

The length of SequenceNumber is 128 bits. As SequenceNumber continues to increase for a particular MessageGroupId.

mD5OfMessageSystemAttributes :: Maybe Text

An MD5 digest of the non-URL-encoded message system attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

mD5OfMessageAttributes :: Maybe Text

An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

id :: Text

An identifier for the message in this batch.

messageId :: Text

An identifier for the message.

mD5OfMessageBody :: Text

An MD5 digest of the non-URL-encoded message body string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

newSendMessageBatchResultEntry

Arguments

:: Text
$sel:id:SendMessageBatchResultEntry'

-> Text
$sel:messageId:SendMessageBatchResultEntry'

-> Text
$sel:mD5OfMessageBody:SendMessageBatchResultEntry'

-> SendMessageBatchResultEntry

Create a value of SendMessageBatchResultEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sequenceNumber:SendMessageBatchResultEntry', sendMessageBatchResultEntry_sequenceNumber - This parameter applies only to FIFO (first-in-first-out) queues.

The large, non-consecutive number that Amazon SQS assigns to each message.

The length of SequenceNumber is 128 bits. As SequenceNumber continues to increase for a particular MessageGroupId.

$sel:mD5OfMessageSystemAttributes:SendMessageBatchResultEntry', sendMessageBatchResultEntry_mD5OfMessageSystemAttributes - An MD5 digest of the non-URL-encoded message system attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

$sel:mD5OfMessageAttributes:SendMessageBatchResultEntry', sendMessageBatchResultEntry_mD5OfMessageAttributes - An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

$sel:id:SendMessageBatchResultEntry', sendMessageBatchResultEntry_id - An identifier for the message in this batch.

$sel:messageId:SendMessageBatchResultEntry', sendMessageBatchResultEntry_messageId - An identifier for the message.

$sel:mD5OfMessageBody:SendMessageBatchResultEntry', sendMessageBatchResultEntry_mD5OfMessageBody - An MD5 digest of the non-URL-encoded message body string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

sendMessageBatchResultEntry_sequenceNumber :: Lens' SendMessageBatchResultEntry (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) queues.

The large, non-consecutive number that Amazon SQS assigns to each message.

The length of SequenceNumber is 128 bits. As SequenceNumber continues to increase for a particular MessageGroupId.

sendMessageBatchResultEntry_mD5OfMessageSystemAttributes :: Lens' SendMessageBatchResultEntry (Maybe Text)

An MD5 digest of the non-URL-encoded message system attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

sendMessageBatchResultEntry_mD5OfMessageAttributes :: Lens' SendMessageBatchResultEntry (Maybe Text)

An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

sendMessageBatchResultEntry_id :: Lens' SendMessageBatchResultEntry Text

An identifier for the message in this batch.

sendMessageBatchResultEntry_messageId :: Lens' SendMessageBatchResultEntry Text

An identifier for the message.

sendMessageBatchResultEntry_mD5OfMessageBody :: Lens' SendMessageBatchResultEntry Text

An MD5 digest of the non-URL-encoded message body string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.