Network.AWS.SQS.Types.BatchResultErrorEntry

Description

Documentation

data BatchResultErrorEntry

Gives a detailed description of the result of an action on each entry in the request.

See: newBatchResultErrorEntry smart constructor.

Constructors

BatchResultErrorEntry'

Fields

message :: Maybe Text

A message explaining why the action failed on this entry.

id :: Text

The Id of an entry in a batch request.

senderFault :: Bool

Specifies whether the error happened due to the caller of the batch API action.

code :: Text

An error code representing why the action failed on this entry.

newBatchResultErrorEntry

Arguments

:: Text
$sel:id:BatchResultErrorEntry'

-> Bool
$sel:senderFault:BatchResultErrorEntry'

-> Text
$sel:code:BatchResultErrorEntry'

-> BatchResultErrorEntry

Create a value of BatchResultErrorEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:BatchResultErrorEntry', batchResultErrorEntry_message - A message explaining why the action failed on this entry.

$sel:id:BatchResultErrorEntry', batchResultErrorEntry_id - The Id of an entry in a batch request.

$sel:senderFault:BatchResultErrorEntry', batchResultErrorEntry_senderFault - Specifies whether the error happened due to the caller of the batch API action.

$sel:code:BatchResultErrorEntry', batchResultErrorEntry_code - An error code representing why the action failed on this entry.

batchResultErrorEntry_message :: Lens' BatchResultErrorEntry (Maybe Text)

A message explaining why the action failed on this entry.

batchResultErrorEntry_id :: Lens' BatchResultErrorEntry Text

The Id of an entry in a batch request.

batchResultErrorEntry_senderFault :: Lens' BatchResultErrorEntry Bool

Specifies whether the error happened due to the caller of the batch API action.

batchResultErrorEntry_code :: Lens' BatchResultErrorEntry Text

An error code representing why the action failed on this entry.