Network.AWS.SQS.Types.SendMessageBatchResultEntry

Description

Documentation

data SendMessageBatchResultEntry

Encloses a MessageId for a successfully-enqueued message in a SendMessageBatch.

See: newSendMessageBatchResultEntry smart constructor.

Constructors

SendMessageBatchResultEntry'

Fields

sequenceNumber :: Maybe Text

This parameter applies only to FIFO (first-in-first-out) queues.

The large, non-consecutive number that Amazon SQS assigns to each message.

The length of SequenceNumber is 128 bits. As SequenceNumber continues to increase for a particular MessageGroupId.

mD5OfMessageSystemAttributes :: Maybe Text

An MD5 digest of the non-URL-encoded message system attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

mD5OfMessageAttributes :: Maybe Text

An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

id :: Text

An identifier for the message in this batch.

messageId :: Text

An identifier for the message.

mD5OfMessageBody :: Text

An MD5 digest of the non-URL-encoded message body string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

newSendMessageBatchResultEntry

Arguments

:: Text
$sel:id:SendMessageBatchResultEntry'

-> Text
$sel:messageId:SendMessageBatchResultEntry'

-> Text
$sel:mD5OfMessageBody:SendMessageBatchResultEntry'

-> SendMessageBatchResultEntry

Create a value of SendMessageBatchResultEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sequenceNumber:SendMessageBatchResultEntry', sendMessageBatchResultEntry_sequenceNumber - This parameter applies only to FIFO (first-in-first-out) queues.

The large, non-consecutive number that Amazon SQS assigns to each message.

The length of SequenceNumber is 128 bits. As SequenceNumber continues to increase for a particular MessageGroupId.

$sel:mD5OfMessageSystemAttributes:SendMessageBatchResultEntry', sendMessageBatchResultEntry_mD5OfMessageSystemAttributes - An MD5 digest of the non-URL-encoded message system attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

$sel:mD5OfMessageAttributes:SendMessageBatchResultEntry', sendMessageBatchResultEntry_mD5OfMessageAttributes - An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

$sel:id:SendMessageBatchResultEntry', sendMessageBatchResultEntry_id - An identifier for the message in this batch.

$sel:messageId:SendMessageBatchResultEntry', sendMessageBatchResultEntry_messageId - An identifier for the message.

$sel:mD5OfMessageBody:SendMessageBatchResultEntry', sendMessageBatchResultEntry_mD5OfMessageBody - An MD5 digest of the non-URL-encoded message body string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

sendMessageBatchResultEntry_sequenceNumber :: Lens' SendMessageBatchResultEntry (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) queues.

The large, non-consecutive number that Amazon SQS assigns to each message.

The length of SequenceNumber is 128 bits. As SequenceNumber continues to increase for a particular MessageGroupId.

sendMessageBatchResultEntry_mD5OfMessageSystemAttributes :: Lens' SendMessageBatchResultEntry (Maybe Text)

An MD5 digest of the non-URL-encoded message system attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

sendMessageBatchResultEntry_mD5OfMessageAttributes :: Lens' SendMessageBatchResultEntry (Maybe Text)

An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

sendMessageBatchResultEntry_id :: Lens' SendMessageBatchResultEntry Text

An identifier for the message in this batch.

sendMessageBatchResultEntry_messageId :: Lens' SendMessageBatchResultEntry Text

An identifier for the message.

sendMessageBatchResultEntry_mD5OfMessageBody :: Lens' SendMessageBatchResultEntry Text

An MD5 digest of the non-URL-encoded message body string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.