Network.AWS.SQS.V2012_11_05.DeleteQueue

Description

Deletes the queue specified by the queue URL, regardless of whether the queue is empty. If the specified queue does not exist, Amazon SQS returns a successful response. Use DeleteQueue with care; once you delete your queue, any messages in the queue are no longer available. When you delete a queue, the deletion process takes up to 60 seconds. Requests you send involving that queue during the 60 seconds might succeed. For example, a SendMessage request might succeed, but after the 60 seconds, the queue and that message you sent no longer exist. Also, when you delete a queue, you must wait at least 60 seconds before creating a queue with the same name. We reserve the right to delete queues that have had no activity for more than 30 days. For more information, see How Amazon SQS Queues Work in the Amazon SQS Developer Guide. The following example Query request deletes the specified queue. http:/sqs.us-east-1.amazonaws.com123456789012testQueue ?Action=DeleteQueue &Version=2009-02-01 &SignatureMethod=HmacSHA256 &Expires=2009-04-18T22%3A52%3A43PST &AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &SignatureVersion=2 &Signature=Dqlp3Sd6ljTUA9Uf6SGtEExwUQEXAMPLE 6fde8d1e-52cd-4581-8cd9-c512f4c64223.

Documentation

data DeleteQueue

Constructors

DeleteQueue

Fields

_dqrQueueUrl :: Text

The URL of the Amazon SQS queue to take action on.

dqrQueueUrl :: Iso' DeleteQueue Text

data DeleteQueueResponse

Constructors

DeleteQueueResponse