Network.AWS.SQS.V2012_11_05.RemovePermission

Description

Revokes any permissions in the queue policy that matches the specified Label parameter. Only the owner of the queue can remove permissions. The following example Query request removes the testLabel permission on the queue named testQueue. http:/sqs.us-east-1.amazonaws.com123456789012testQueue ?Action=RemovePermission &Label=testLabel &Version=2009-02-01 &SignatureMethod=HmacSHA256 &Expires=2009-04-18T22%3A52%3A43PST &AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &SignatureVersion=2 &Signature=Dqlp3Sd6ljTUA9Uf6SGtEExwUQEXAMPLE f8bdb362-6616-42c0-977a-ce9a8bcce3bb.

Documentation

data RemovePermission

Constructors

RemovePermission

Fields

_rprQueueUrl :: Text

The URL of the Amazon SQS queue to take action on.

_rprLabel :: Text

The identification of the permission to remove. This is the label added with the AddPermission action.

rprQueueUrl :: Lens' RemovePermission Text

rprLabel :: Lens' RemovePermission Text

data RemovePermissionResponse

Constructors

RemovePermissionResponse