Network.AWS.SQS.V2012_11_05.SendMessage

Description

Delivers a message to the specified queue. With Amazon SQS, you now have the ability to send large payload messages that are up to 256KB (262,144 bytes) in size. To send large payloads, you must use an AWS SDK that supports SigV4 signing. To verify whether SigV4 is supported for an AWS SDK, check the SDK release notes. The following list shows the characters (in Unicode) allowed in your message, according to the W3C XML specification. For more information, go to http:/www.w3.orgTRREC-xml#charsets If you send any characters not included in the list, your request will be rejected. #x9 | #xA | #xD | [#x20 to #xD7FF] | [#xE000 to #xFFFD] | [#x10000 to #x10FFFF] The following example SendMessage request sends a message containing "This is a test message" to the queue. You must URL encode the entire URL; however, we've URL encoded only the message body to make the example easier for you to read. The following example response includes the MD5 digest for "This is a test message". http:/sqs.us-east-1.amazonaws.com123456789012testQueue ?Action=SendMessage &MessageBody=This+is+a+test+message &MessageAttribute.1.Name=test_attribute_name_1 &MessageAttribute.1.Value.StringValue=test_attribute_value_1 &MessageAttribute.1.Value.DataType=String &MessageAttribute.2.Name=test_attribute_name_2 &MessageAttribute.2.Value.StringValue=test_attribute_value_2 &MessageAttribute.2.Value.DataType=String &Version=2012-11-05 &SignatureMethod=HmacSHA256 &Expires=2014-05-05T22%3A52%3A43PST &AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &SignatureVersion=2 &Signature=Dqlp3Sd6ljTUA9Uf6SGtEExwUQEXAMPLE fafb00f5732ab283681e124bf8747ed1 3ae8f24a165a8cedc005670c81a27295 5fea7756-0ea4-451a-a703-a558b933e274 27daac76-34dd-47df-bd01-1f6e873584a0.

Documentation

sendMessage

Arguments

:: Text
_smrQueueUrl

-> Text
_smrMessageBody

-> SendMessage

Minimum specification for a SendMessage request.

data SendMessage

Constructors

SendMessage

Fields

_smrQueueUrl :: Text

The URL of the Amazon SQS queue to take action on.

_smrMessageBody :: Text

The message to send. String maximum 256 KB in size. For a list of allowed characters, see the preceding important note.

_smrDelaySeconds :: Maybe Integer

The number of seconds (0 to 900 - 15 minutes) to delay a specific message. Messages with a positive DelaySeconds value become available for processing after the delay time is finished. If you don't specify a value, the default value for the queue applies.

_smrMessageAttributes :: HashMap Text MessageAttributeValue

Each message attribute consists of a Name, Type, and Value. For more information, see Message Attribute Items.

smrQueueUrl :: Lens' SendMessage Text

smrMessageBody :: Lens' SendMessage Text

smrMessageAttributes :: Lens' SendMessage (HashMap Text MessageAttributeValue)

smrDelaySeconds :: Lens' SendMessage (Maybe Integer)

data SendMessageResponse

Constructors

SendMessageResponse

Fields

_smsMessageId :: Maybe Text

An element containing the message ID of the message sent to the queue. For more information, see Queue and Message Identifiers in the Amazon SQS Developer Guide.

_smsMD5OfMessageBody :: Maybe Text

An MD5 digest of the non-URL-encoded message body string. This can be used to verify that Amazon SQS received the message correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http:/www.faqs.orgrfcs/rfc1321.html.

_smsMD5OfMessageAttributes :: Maybe Text

An MD5 digest of the non-URL-encoded message attribute string. This can be used to verify that Amazon SQS received the message correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http:/www.faqs.orgrfcs/rfc1321.html.

smsMessageId :: Lens' SendMessageResponse (Maybe Text)

smsMD5OfMessageBody :: Lens' SendMessageResponse (Maybe Text)

smsMD5OfMessageAttributes :: Lens' SendMessageResponse (Maybe Text)