Network.AWS.SQS.V2012_11_05.Types

Description

Amazon Simple Queue Service (SQS) is a fast, reliable, scalable, fully managed message queuing service. SQS makes it simple and cost-effective to decouple the components of a cloud application. You can use SQS to transmit any volume of data, at any level of throughput, without losing messages or requiring other services to be always available. With SQS, you can offload the administrative burden of operating and scaling a highly available messaging cluster, while paying a low price for only what you use.

Documentation

data SQS

Supported version (2012-11-05) of the Amazon Simple Queue Service service.

xmlOptions :: Tagged a XMLOptions

data QueueAttributeName

Constructors

QueueAttributeNameApproximateNumberOfMessages
ApproximateNumberOfMessages

QueueAttributeNameApproximateNumberOfMessagesDelayed
ApproximateNumberOfMessagesDelayed

QueueAttributeNameApproximateNumberOfMessagesNotVisible
ApproximateNumberOfMessagesNotVisible

QueueAttributeNameCreatedTimestamp
CreatedTimestamp

QueueAttributeNameDelaySeconds
DelaySeconds

QueueAttributeNameLastModifiedTimestamp
LastModifiedTimestamp

QueueAttributeNameMaximumMessageSize
MaximumMessageSize

QueueAttributeNameMessageRetentionPeriod
MessageRetentionPeriod

QueueAttributeNamePolicy
Policy

QueueAttributeNameQueueArn
QueueArn

QueueAttributeNameReceiveMessageWaitTimeSeconds
ReceiveMessageWaitTimeSeconds

QueueAttributeNameRedrivePolicy
RedrivePolicy

QueueAttributeNameVisibilityTimeout
VisibilityTimeout

newtype ChangeMessageVisibilityBatchResultEntry

Encloses the id of an entry in ChangeMessageVisibilityBatch.

Constructors

ChangeMessageVisibilityBatchResultEntry

Fields

_cmvbrfId :: Text

Represents a message whose visibility timeout has been changed successfully.

newtype DeleteMessageBatchResultEntry

Encloses the id an entry in DeleteMessageBatch.

Constructors

DeleteMessageBatchResultEntry

Fields

_dmbrfId :: Text

Represents a successfully deleted message.

data BatchResultErrorEntry

This is used in the responses of batch API to give a detailed description of the result of an action on each entry in the request.

Constructors

BatchResultErrorEntry

Fields

_breeSenderFault :: Bool

Whether the error happened due to the sender's fault.

_breeId :: Text

The id of an entry in a batch request.

_breeCode :: Text

An error code representing why the action failed on this entry.

_breeMessage :: Maybe Text

A message explaining why the action failed on this entry.

data ChangeMessageVisibilityBatchRequestEntry

Encloses a receipt handle and an entry id for each message in ChangeMessageVisibilityBatch. All of the following parameters are list parameters that must be prefixed with ChangeMessageVisibilityBatchRequestEntry.n, where n is an integer value starting with 1. For example, a parameter list for this action might look like this: &amp;ChangeMessageVisibilityBatchRequestEntry.1.Id=change_visibility_msg_2 &amp;ChangeMessageVisibilityBatchRequestEntry.1.ReceiptHandle=Your_Receipt_Handle &amp;ChangeMessageVisibilityBatchRequestEntry.1.VisibilityTimeout=45.

Constructors

ChangeMessageVisibilityBatchRequestEntry

Fields

_cmvbreVisibilityTimeout :: Maybe Integer

The new value (in seconds) for the message's visibility timeout.

_cmvbreId :: Text

An identifier for this particular receipt handle. This is used to communicate the result. Note that the Ids of a batch request need to be unique within the request.

_cmvbreReceiptHandle :: Text

A receipt handle.

data DeleteMessageBatchRequestEntry

Encloses a receipt handle and an identifier for it.

Constructors

DeleteMessageBatchRequestEntry

Fields

_dmbreId :: Text

An identifier for this particular receipt handle. This is used to communicate the result. Note that the Ids of a batch request need to be unique within the request.

_dmbreReceiptHandle :: Text

A receipt handle.

data Message

An Amazon SQS message.

Constructors

Message

Fields

_nMessageAttributes :: HashMap Text MessageAttributeValue

Each message attribute consists of a Name, Type, and Value. For more information, see Message Attribute Items.

_nMD5OfBody :: Maybe Text

An MD5 digest of the non-URL-encoded message body string.

_nBody :: Maybe Text

The message's contents (not URL-encoded).

_nAttributes :: HashMap QueueAttributeName Text

SenderId, SentTimestamp, ApproximateReceiveCount, and/or ApproximateFirstReceiveTimestamp. SentTimestamp and ApproximateFirstReceiveTimestamp are each returned as an integer representing the epoch time in milliseconds.

_nReceiptHandle :: Maybe Text

An identifier associated with the act of receiving the message. A new receipt handle is returned every time you receive a message. When deleting a message, you provide the last received receipt handle to delete the message.

_nMessageId :: Maybe Text

A unique identifier for the message. Message IDs are considered unique across all AWS accounts for an extended period of time.

_nMD5OfMessageAttributes :: Maybe Text

An MD5 digest of the non-URL-encoded message attribute string. This can be used to verify that Amazon SQS received the message correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http:/www.faqs.orgrfcs/rfc1321.html.

data MessageAttributeValue

The user-specified message attribute value. For string data types, the value attribute has the same restrictions on the content as the message body. For more information, see SendMessage. Name, type, and value must not be empty or null. In addition, the message body should not be empty or null. All parts of the message attribute, including name, type, and value, are included in the message size restriction, which is currently 256 KB (262,144 bytes).

Constructors

MessageAttributeValue

Fields

_mavBinaryValue :: Maybe ByteString

Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.

_mavStringListValues :: [Text]

Not implemented. Reserved for future use.

_mavStringValue :: Maybe Text

Strings are Unicode with UTF8 binary encoding. For a list of code values, see http:/en.wikipedia.orgwiki/ASCII#ASCII_printable_characters.

_mavBinaryListValues :: [ByteString]

Not implemented. Reserved for future use.

_mavDataType :: Text

Amazon SQS supports the following logical data types: String, Number, and Binary. In addition, you can append your own custom labels. For more information, see Message Attribute Data Types.

data SendMessageBatchRequestEntry

Contains the details of a single Amazon SQS message along with a Id.

Constructors

SendMessageBatchRequestEntry

Fields

_smbreMessageAttributes :: HashMap Text MessageAttributeValue

Each message attribute consists of a Name, Type, and Value. For more information, see Message Attribute Items.

_smbreDelaySeconds :: Maybe Integer

The number of seconds for which the message has to be delayed.

_smbreId :: Text

An identifier for the message in this batch. This is used to communicate the result. Note that the Ids of a batch request need to be unique within the request.

_smbreMessageBody :: Text

Body of the message.

data SendMessageBatchResultEntry

Encloses a message ID for successfully enqueued message of a SendMessageBatch.

Constructors

SendMessageBatchResultEntry

Fields

_smbrfId :: Text

An identifier for the message in this batch.

_smbrfMessageId :: Text

An identifier for the message.

_smbrfMD5OfMessageBody :: Text

An MD5 digest of the non-URL-encoded message body string. This can be used to verify that Amazon SQS received the message correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http:/www.faqs.orgrfcs/rfc1321.html.

_smbrfMD5OfMessageAttributes :: Maybe Text

An MD5 digest of the non-URL-encoded message attribute string. This can be used to verify that Amazon SQS received the message batch correctly. Amazon SQS first URL decodes the message before creating the MD5 digest. For information about MD5, go to http:/www.faqs.orgrfcs/rfc1321.html.

cmvbrfId :: Iso' ChangeMessageVisibilityBatchResultEntry Text

changeMessageVisibilityBatchResultEntry :: Iso' Text ChangeMessageVisibilityBatchResultEntry

dmbrfId :: Iso' DeleteMessageBatchResultEntry Text

deleteMessageBatchResultEntry :: Iso' Text DeleteMessageBatchResultEntry

breeSenderFault :: Lens' BatchResultErrorEntry Bool

breeMessage :: Lens' BatchResultErrorEntry (Maybe Text)

breeId :: Lens' BatchResultErrorEntry Text

breeCode :: Lens' BatchResultErrorEntry Text

cmvbreVisibilityTimeout :: Lens' ChangeMessageVisibilityBatchRequestEntry (Maybe Integer)

cmvbreReceiptHandle :: Lens' ChangeMessageVisibilityBatchRequestEntry Text

cmvbreId :: Lens' ChangeMessageVisibilityBatchRequestEntry Text

dmbreReceiptHandle :: Lens' DeleteMessageBatchRequestEntry Text

dmbreId :: Lens' DeleteMessageBatchRequestEntry Text

nReceiptHandle :: Lens' Message (Maybe Text)

nMessageId :: Lens' Message (Maybe Text)

nMessageAttributes :: Lens' Message (HashMap Text MessageAttributeValue)

nMD5OfMessageAttributes :: Lens' Message (Maybe Text)

nMD5OfBody :: Lens' Message (Maybe Text)

nBody :: Lens' Message (Maybe Text)

nAttributes :: Lens' Message (HashMap QueueAttributeName Text)

mavStringValue :: Lens' MessageAttributeValue (Maybe Text)

mavStringListValues :: Lens' MessageAttributeValue [Text]

mavDataType :: Lens' MessageAttributeValue Text

mavBinaryValue :: Lens' MessageAttributeValue (Maybe ByteString)

mavBinaryListValues :: Lens' MessageAttributeValue [ByteString]

smbreMessageBody :: Lens' SendMessageBatchRequestEntry Text

smbreMessageAttributes :: Lens' SendMessageBatchRequestEntry (HashMap Text MessageAttributeValue)

smbreId :: Lens' SendMessageBatchRequestEntry Text

smbreDelaySeconds :: Lens' SendMessageBatchRequestEntry (Maybe Integer)

smbrfMessageId :: Lens' SendMessageBatchResultEntry Text

smbrfMD5OfMessageBody :: Lens' SendMessageBatchResultEntry Text

smbrfMD5OfMessageAttributes :: Lens' SendMessageBatchResultEntry (Maybe Text)

smbrfId :: Lens' SendMessageBatchResultEntry Text