Network.AWS.SSM.AssociateOpsItemRelatedItem

Description

Associates a related item to a Systems Manager OpsCenter OpsItem. For example, you can associate an Incident Manager incident or analysis with an OpsItem. Incident Manager and OpsCenter are capabilities of Amazon Web Services Systems Manager.

Creating a Request

data AssociateOpsItemRelatedItem

See: newAssociateOpsItemRelatedItem smart constructor.

Constructors

AssociateOpsItemRelatedItem'

Fields

opsItemId :: Text

The ID of the OpsItem to which you want to associate a resource as a related item.

associationType :: Text

The type of association that you want to create between an OpsItem and a resource. OpsCenter supports IsParentOf and RelatesTo association types.

resourceType :: Text

The type of resource that you want to associate with an OpsItem. OpsCenter supports the following types:

AWS::SSMIncidents::IncidentRecord: an Incident Manager incident.

AWS::SSM::Document: a Systems Manager (SSM) document.

resourceUri :: Text

The Amazon Resource Name (ARN) of the Amazon Web Services resource that you want to associate with the OpsItem.

newAssociateOpsItemRelatedItem

Arguments

:: Text
$sel:opsItemId:AssociateOpsItemRelatedItem'

-> Text
$sel:associationType:AssociateOpsItemRelatedItem'

-> Text
$sel:resourceType:AssociateOpsItemRelatedItem'

-> Text
$sel:resourceUri:AssociateOpsItemRelatedItem'

-> AssociateOpsItemRelatedItem

Create a value of AssociateOpsItemRelatedItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:opsItemId:AssociateOpsItemRelatedItem', associateOpsItemRelatedItem_opsItemId - The ID of the OpsItem to which you want to associate a resource as a related item.

$sel:associationType:AssociateOpsItemRelatedItem', associateOpsItemRelatedItem_associationType - The type of association that you want to create between an OpsItem and a resource. OpsCenter supports IsParentOf and RelatesTo association types.

$sel:resourceType:AssociateOpsItemRelatedItem', associateOpsItemRelatedItem_resourceType - The type of resource that you want to associate with an OpsItem. OpsCenter supports the following types:

AWS::SSMIncidents::IncidentRecord: an Incident Manager incident.

AWS::SSM::Document: a Systems Manager (SSM) document.

$sel:resourceUri:AssociateOpsItemRelatedItem', associateOpsItemRelatedItem_resourceUri - The Amazon Resource Name (ARN) of the Amazon Web Services resource that you want to associate with the OpsItem.

Request Lenses

associateOpsItemRelatedItem_opsItemId :: Lens' AssociateOpsItemRelatedItem Text

The ID of the OpsItem to which you want to associate a resource as a related item.

associateOpsItemRelatedItem_associationType :: Lens' AssociateOpsItemRelatedItem Text

The type of association that you want to create between an OpsItem and a resource. OpsCenter supports IsParentOf and RelatesTo association types.

associateOpsItemRelatedItem_resourceType :: Lens' AssociateOpsItemRelatedItem Text

The type of resource that you want to associate with an OpsItem. OpsCenter supports the following types:

AWS::SSMIncidents::IncidentRecord: an Incident Manager incident.

AWS::SSM::Document: a Systems Manager (SSM) document.

associateOpsItemRelatedItem_resourceUri :: Lens' AssociateOpsItemRelatedItem Text

The Amazon Resource Name (ARN) of the Amazon Web Services resource that you want to associate with the OpsItem.

Destructuring the Response

data AssociateOpsItemRelatedItemResponse

See: newAssociateOpsItemRelatedItemResponse smart constructor.

Constructors

AssociateOpsItemRelatedItemResponse'

Fields

associationId :: Maybe Text

The association ID.

httpStatus :: Int

The response's http status code.

newAssociateOpsItemRelatedItemResponse

Arguments

:: Int
$sel:httpStatus:AssociateOpsItemRelatedItemResponse'

-> AssociateOpsItemRelatedItemResponse

Create a value of AssociateOpsItemRelatedItemResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associationId:AssociateOpsItemRelatedItemResponse', associateOpsItemRelatedItemResponse_associationId - The association ID.

$sel:httpStatus:AssociateOpsItemRelatedItemResponse', associateOpsItemRelatedItemResponse_httpStatus - The response's http status code.

Response Lenses

associateOpsItemRelatedItemResponse_associationId :: Lens' AssociateOpsItemRelatedItemResponse (Maybe Text)

The association ID.

associateOpsItemRelatedItemResponse_httpStatus :: Lens' AssociateOpsItemRelatedItemResponse Int

The response's http status code.