Network.AWS.SSM.CreateActivation

Description

Generates an activation code and activation ID you can use to register your on-premises server or virtual machine (VM) with Amazon Web Services Systems Manager. Registering these machines with Systems Manager makes it possible to manage them using Systems Manager capabilities. You use the activation code and ID when installing SSM Agent on machines in your hybrid environment. For more information about requirements for managing on-premises instances and VMs using Systems Manager, see Setting up Amazon Web Services Systems Manager for hybrid environments in the Amazon Web Services Systems Manager User Guide.

On-premises servers or VMs that are registered with Systems Manager and Amazon Elastic Compute Cloud (Amazon EC2) instances that you manage with Systems Manager are all called managed instances.

Creating a Request

data CreateActivation

See: newCreateActivation smart constructor.

Constructors

CreateActivation'

Fields

registrationLimit :: Maybe Natural

Specify the maximum number of managed instances you want to register. The default value is 1.

defaultInstanceName :: Maybe Text

The name of the registered, managed instance as it will appear in the Amazon Web Services Systems Manager console or when you use the Amazon Web Services command line tools to list Systems Manager resources.

Don't enter personally identifiable information in this field.

expirationDate :: Maybe POSIX

The date by which this activation request should expire, in timestamp format, such as "2021-07-07T00:00:00". You can specify a date up to 30 days in advance. If you don't provide an expiration date, the activation code expires in 24 hours.

tags :: Maybe [Tag]

Optional metadata that you assign to a resource. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag an activation to identify which servers or virtual machines (VMs) in your on-premises environment you intend to activate. In this case, you could specify the following key-value pairs:

Key=OS,Value=Windows

Key=Environment,Value=Production

When you install SSM Agent on your on-premises servers and VMs, you specify an activation ID and code. When you specify the activation ID and code, tags assigned to the activation are automatically applied to the on-premises servers or VMs.

You can't add tags to or delete tags from an existing activation. You can tag your on-premises servers and VMs after they connect to Systems Manager for the first time and are assigned a managed instance ID. This means they are listed in the Amazon Web Services Systems Manager console with an ID that is prefixed with "mi-". For information about how to add tags to your managed instances, see AddTagsToResource. For information about how to remove tags from your managed instances, see RemoveTagsFromResource.

description :: Maybe Text

A user-defined description of the resource that you want to register with Systems Manager.

Don't enter personally identifiable information in this field.

iamRole :: Text

The name of the Identity and Access Management (IAM) role that you want to assign to the managed instance. This IAM role must provide AssumeRole permissions for the Amazon Web Services Systems Manager service principal ssm.amazonaws.com. For more information, see Create an IAM service role for a hybrid environment in the Amazon Web Services Systems Manager User Guide.

newCreateActivation

Arguments

:: Text
$sel:iamRole:CreateActivation'

-> CreateActivation

Create a value of CreateActivation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registrationLimit:CreateActivation', createActivation_registrationLimit - Specify the maximum number of managed instances you want to register. The default value is 1.

$sel:defaultInstanceName:CreateActivation', createActivation_defaultInstanceName - The name of the registered, managed instance as it will appear in the Amazon Web Services Systems Manager console or when you use the Amazon Web Services command line tools to list Systems Manager resources.

Don't enter personally identifiable information in this field.

$sel:expirationDate:CreateActivation', createActivation_expirationDate - The date by which this activation request should expire, in timestamp format, such as "2021-07-07T00:00:00". You can specify a date up to 30 days in advance. If you don't provide an expiration date, the activation code expires in 24 hours.

$sel:tags:CreateActivation', createActivation_tags - Optional metadata that you assign to a resource. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag an activation to identify which servers or virtual machines (VMs) in your on-premises environment you intend to activate. In this case, you could specify the following key-value pairs:

Key=OS,Value=Windows

Key=Environment,Value=Production

When you install SSM Agent on your on-premises servers and VMs, you specify an activation ID and code. When you specify the activation ID and code, tags assigned to the activation are automatically applied to the on-premises servers or VMs.

You can't add tags to or delete tags from an existing activation. You can tag your on-premises servers and VMs after they connect to Systems Manager for the first time and are assigned a managed instance ID. This means they are listed in the Amazon Web Services Systems Manager console with an ID that is prefixed with "mi-". For information about how to add tags to your managed instances, see AddTagsToResource. For information about how to remove tags from your managed instances, see RemoveTagsFromResource.

$sel:description:CreateActivation', createActivation_description - A user-defined description of the resource that you want to register with Systems Manager.

Don't enter personally identifiable information in this field.

$sel:iamRole:CreateActivation', createActivation_iamRole - The name of the Identity and Access Management (IAM) role that you want to assign to the managed instance. This IAM role must provide AssumeRole permissions for the Amazon Web Services Systems Manager service principal ssm.amazonaws.com. For more information, see Create an IAM service role for a hybrid environment in the Amazon Web Services Systems Manager User Guide.

Request Lenses

createActivation_registrationLimit :: Lens' CreateActivation (Maybe Natural)

Specify the maximum number of managed instances you want to register. The default value is 1.

createActivation_defaultInstanceName :: Lens' CreateActivation (Maybe Text)

The name of the registered, managed instance as it will appear in the Amazon Web Services Systems Manager console or when you use the Amazon Web Services command line tools to list Systems Manager resources.

Don't enter personally identifiable information in this field.

createActivation_expirationDate :: Lens' CreateActivation (Maybe UTCTime)

The date by which this activation request should expire, in timestamp format, such as "2021-07-07T00:00:00". You can specify a date up to 30 days in advance. If you don't provide an expiration date, the activation code expires in 24 hours.

createActivation_tags :: Lens' CreateActivation (Maybe [Tag])

Optional metadata that you assign to a resource. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag an activation to identify which servers or virtual machines (VMs) in your on-premises environment you intend to activate. In this case, you could specify the following key-value pairs:

Key=OS,Value=Windows

Key=Environment,Value=Production

When you install SSM Agent on your on-premises servers and VMs, you specify an activation ID and code. When you specify the activation ID and code, tags assigned to the activation are automatically applied to the on-premises servers or VMs.

You can't add tags to or delete tags from an existing activation. You can tag your on-premises servers and VMs after they connect to Systems Manager for the first time and are assigned a managed instance ID. This means they are listed in the Amazon Web Services Systems Manager console with an ID that is prefixed with "mi-". For information about how to add tags to your managed instances, see AddTagsToResource. For information about how to remove tags from your managed instances, see RemoveTagsFromResource.

createActivation_description :: Lens' CreateActivation (Maybe Text)

A user-defined description of the resource that you want to register with Systems Manager.

Don't enter personally identifiable information in this field.

createActivation_iamRole :: Lens' CreateActivation Text

The name of the Identity and Access Management (IAM) role that you want to assign to the managed instance. This IAM role must provide AssumeRole permissions for the Amazon Web Services Systems Manager service principal ssm.amazonaws.com. For more information, see Create an IAM service role for a hybrid environment in the Amazon Web Services Systems Manager User Guide.

Destructuring the Response

data CreateActivationResponse

See: newCreateActivationResponse smart constructor.

Constructors

CreateActivationResponse'

Fields

activationId :: Maybe Text

The ID number generated by the system when it processed the activation. The activation ID functions like a user name.

activationCode :: Maybe Text

The code the system generates when it processes the activation. The activation code functions like a password to validate the activation ID.

httpStatus :: Int

The response's http status code.

newCreateActivationResponse

Arguments

:: Int
$sel:httpStatus:CreateActivationResponse'

-> CreateActivationResponse

Create a value of CreateActivationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:activationId:CreateActivationResponse', createActivationResponse_activationId - The ID number generated by the system when it processed the activation. The activation ID functions like a user name.

$sel:activationCode:CreateActivationResponse', createActivationResponse_activationCode - The code the system generates when it processes the activation. The activation code functions like a password to validate the activation ID.

$sel:httpStatus:CreateActivationResponse', createActivationResponse_httpStatus - The response's http status code.

Response Lenses

createActivationResponse_activationId :: Lens' CreateActivationResponse (Maybe Text)

The ID number generated by the system when it processed the activation. The activation ID functions like a user name.

createActivationResponse_activationCode :: Lens' CreateActivationResponse (Maybe Text)

The code the system generates when it processes the activation. The activation code functions like a password to validate the activation ID.

createActivationResponse_httpStatus :: Lens' CreateActivationResponse Int

The response's http status code.