Network.AWS.SSM.CreateAssociation

Description

A State Manager association defines the state that you want to maintain on your instances. For example, an association can specify that anti-virus software must be installed and running on your instances, or that certain ports must be closed. For static targets, the association specifies a schedule for when the configuration is reapplied. For dynamic targets, such as an Amazon Web Services resource group or an Amazon Web Services autoscaling group, State Manager, a capability of Amazon Web Services Systems Manager applies the configuration when new instances are added to the group. The association also specifies actions to take when applying the configuration. For example, an association for anti-virus software might run once a day. If the software isn't installed, then State Manager installs it. If the software is installed, but the service isn't running, then the association might instruct State Manager to start the service.

Creating a Request

data CreateAssociation

See: newCreateAssociation smart constructor.

Constructors

CreateAssociation'

Fields

maxErrors :: Maybe Text

The number of errors that are allowed before the system stops sending requests to run the association on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops sending requests when the fourth error is received. If you specify 0, then the system stops sending requests after the first error is returned. If you run an association on 50 instances and set MaxError to 10%, then the system stops sending the request when the sixth error is received.

Executions that are already running an association when MaxErrors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set MaxConcurrency to 1 so that executions proceed one at a time.

instanceId :: Maybe Text

The instance ID.

InstanceId has been deprecated. To specify an instance ID for an association, use the Targets parameter. Requests that include the parameter InstanceID with Systems Manager documents (SSM documents) that use schema version 2.0 or later will fail. In addition, if you use the parameter InstanceId, you can't use the parameters AssociationName, DocumentVersion, MaxErrors, MaxConcurrency, OutputLocation, or ScheduleExpression. To use these parameters, you must use the Targets parameter.

complianceSeverity :: Maybe AssociationComplianceSeverity

The severity level to assign to the association.

automationTargetParameterName :: Maybe Text

Specify the target for the association. This target is required for associations that use an Automation runbook and target resources by using rate controls. Automation is a capability of Amazon Web Services Systems Manager.

targets :: Maybe [Target]

The targets for the association. You can target instances by using tags, Amazon Web Services resource groups, all instances in an Amazon Web Services account, or individual instance IDs. For more information about choosing targets for an association, see Using targets and rate controls with State Manager associations in the Amazon Web Services Systems Manager User Guide.

scheduleExpression :: Maybe Text

A cron expression when the association will be applied to the target(s).

targetLocations :: Maybe (NonEmpty TargetLocation)

A location is a combination of Amazon Web Services Regions and Amazon Web Services accounts where you want to run the association. Use this action to create an association in multiple Regions and multiple accounts.

maxConcurrency :: Maybe Text

The maximum number of targets allowed to run the association at the same time. You can specify a number, for example 10, or a percentage of the target set, for example 10%. The default value is 100%, which means all targets run the association at the same time.

If a new instance starts and attempts to run an association while Systems Manager is running MaxConcurrency associations, the association is allowed to run. During the next association interval, the new instance will process its association within the limit specified for MaxConcurrency.

calendarNames :: Maybe [Text]

The names or Amazon Resource Names (ARNs) of the Change Calendar type documents you want to gate your associations under. The associations only run when that change calendar is open. For more information, see Amazon Web Services Systems Manager Change Calendar.

associationName :: Maybe Text

Specify a descriptive name for the association.

documentVersion :: Maybe Text

The document version you want to associate with the target(s). Can be a specific version or the default version.

parameters :: Maybe (HashMap Text [Text])

The parameters for the runtime configuration of the document.

syncCompliance :: Maybe AssociationSyncCompliance

The mode for generating association compliance. You can specify AUTO or MANUAL. In AUTO mode, the system uses the status of the association execution to determine the compliance status. If the association execution runs successfully, then the association is COMPLIANT. If the association execution doesn't run successfully, the association is NON-COMPLIANT.

In MANUAL mode, you must specify the AssociationId as a parameter for the PutComplianceItems API operation. In this case, compliance data isn't managed by State Manager. It is managed by your direct call to the PutComplianceItems API operation.

By default, all associations use AUTO mode.

outputLocation :: Maybe InstanceAssociationOutputLocation

An Amazon Simple Storage Service (Amazon S3) bucket where you want to store the output details of the request.

applyOnlyAtCronInterval :: Maybe Bool

By default, when you create a new association, the system runs it immediately after it is created and then according to the schedule you specified. Specify this option if you don't want an association to run immediately after you create it. This parameter isn't supported for rate expressions.

name :: Text

The name of the SSM Command document or Automation runbook that contains the configuration information for the instance.

You can specify Amazon Web Services-predefined documents, documents you created, or a document that is shared with you from another account.

For Systems Manager documents (SSM documents) that are shared with you from other Amazon Web Services accounts, you must specify the complete SSM document ARN, in the following format:

arn:partition:ssm:region:account-id:document/document-name

For example:

arn:aws:ssm:us-east-2:12345678912:document/My-Shared-Document

For Amazon Web Services-predefined documents and SSM documents you created in your account, you only need to specify the document name. For example, AWS-ApplyPatchBaseline or My-Document.

newCreateAssociation

Arguments

:: Text
$sel:name:CreateAssociation'

-> CreateAssociation

Create a value of CreateAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxErrors:CreateAssociation', createAssociation_maxErrors - The number of errors that are allowed before the system stops sending requests to run the association on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops sending requests when the fourth error is received. If you specify 0, then the system stops sending requests after the first error is returned. If you run an association on 50 instances and set MaxError to 10%, then the system stops sending the request when the sixth error is received.

Executions that are already running an association when MaxErrors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set MaxConcurrency to 1 so that executions proceed one at a time.

$sel:instanceId:CreateAssociation', createAssociation_instanceId - The instance ID.

InstanceId has been deprecated. To specify an instance ID for an association, use the Targets parameter. Requests that include the parameter InstanceID with Systems Manager documents (SSM documents) that use schema version 2.0 or later will fail. In addition, if you use the parameter InstanceId, you can't use the parameters AssociationName, DocumentVersion, MaxErrors, MaxConcurrency, OutputLocation, or ScheduleExpression. To use these parameters, you must use the Targets parameter.

$sel:complianceSeverity:CreateAssociation', createAssociation_complianceSeverity - The severity level to assign to the association.

$sel:automationTargetParameterName:CreateAssociation', createAssociation_automationTargetParameterName - Specify the target for the association. This target is required for associations that use an Automation runbook and target resources by using rate controls. Automation is a capability of Amazon Web Services Systems Manager.

$sel:targets:CreateAssociation', createAssociation_targets - The targets for the association. You can target instances by using tags, Amazon Web Services resource groups, all instances in an Amazon Web Services account, or individual instance IDs. For more information about choosing targets for an association, see Using targets and rate controls with State Manager associations in the Amazon Web Services Systems Manager User Guide.

$sel:scheduleExpression:CreateAssociation', createAssociation_scheduleExpression - A cron expression when the association will be applied to the target(s).

$sel:targetLocations:CreateAssociation', createAssociation_targetLocations - A location is a combination of Amazon Web Services Regions and Amazon Web Services accounts where you want to run the association. Use this action to create an association in multiple Regions and multiple accounts.

$sel:maxConcurrency:CreateAssociation', createAssociation_maxConcurrency - The maximum number of targets allowed to run the association at the same time. You can specify a number, for example 10, or a percentage of the target set, for example 10%. The default value is 100%, which means all targets run the association at the same time.

If a new instance starts and attempts to run an association while Systems Manager is running MaxConcurrency associations, the association is allowed to run. During the next association interval, the new instance will process its association within the limit specified for MaxConcurrency.

$sel:calendarNames:CreateAssociation', createAssociation_calendarNames - The names or Amazon Resource Names (ARNs) of the Change Calendar type documents you want to gate your associations under. The associations only run when that change calendar is open. For more information, see Amazon Web Services Systems Manager Change Calendar.

$sel:associationName:CreateAssociation', createAssociation_associationName - Specify a descriptive name for the association.

$sel:documentVersion:CreateAssociation', createAssociation_documentVersion - The document version you want to associate with the target(s). Can be a specific version or the default version.

$sel:parameters:CreateAssociation', createAssociation_parameters - The parameters for the runtime configuration of the document.

$sel:syncCompliance:CreateAssociation', createAssociation_syncCompliance - The mode for generating association compliance. You can specify AUTO or MANUAL. In AUTO mode, the system uses the status of the association execution to determine the compliance status. If the association execution runs successfully, then the association is COMPLIANT. If the association execution doesn't run successfully, the association is NON-COMPLIANT.

In MANUAL mode, you must specify the AssociationId as a parameter for the PutComplianceItems API operation. In this case, compliance data isn't managed by State Manager. It is managed by your direct call to the PutComplianceItems API operation.

By default, all associations use AUTO mode.

$sel:outputLocation:CreateAssociation', createAssociation_outputLocation - An Amazon Simple Storage Service (Amazon S3) bucket where you want to store the output details of the request.

$sel:applyOnlyAtCronInterval:CreateAssociation', createAssociation_applyOnlyAtCronInterval - By default, when you create a new association, the system runs it immediately after it is created and then according to the schedule you specified. Specify this option if you don't want an association to run immediately after you create it. This parameter isn't supported for rate expressions.

$sel:name:CreateAssociation', createAssociation_name - The name of the SSM Command document or Automation runbook that contains the configuration information for the instance.

You can specify Amazon Web Services-predefined documents, documents you created, or a document that is shared with you from another account.

For Systems Manager documents (SSM documents) that are shared with you from other Amazon Web Services accounts, you must specify the complete SSM document ARN, in the following format:

arn:partition:ssm:region:account-id:document/document-name

For example:

arn:aws:ssm:us-east-2:12345678912:document/My-Shared-Document

For Amazon Web Services-predefined documents and SSM documents you created in your account, you only need to specify the document name. For example, AWS-ApplyPatchBaseline or My-Document.

Request Lenses

createAssociation_maxErrors :: Lens' CreateAssociation (Maybe Text)

The number of errors that are allowed before the system stops sending requests to run the association on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops sending requests when the fourth error is received. If you specify 0, then the system stops sending requests after the first error is returned. If you run an association on 50 instances and set MaxError to 10%, then the system stops sending the request when the sixth error is received.

Executions that are already running an association when MaxErrors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set MaxConcurrency to 1 so that executions proceed one at a time.

createAssociation_instanceId :: Lens' CreateAssociation (Maybe Text)

The instance ID.

InstanceId has been deprecated. To specify an instance ID for an association, use the Targets parameter. Requests that include the parameter InstanceID with Systems Manager documents (SSM documents) that use schema version 2.0 or later will fail. In addition, if you use the parameter InstanceId, you can't use the parameters AssociationName, DocumentVersion, MaxErrors, MaxConcurrency, OutputLocation, or ScheduleExpression. To use these parameters, you must use the Targets parameter.

createAssociation_complianceSeverity :: Lens' CreateAssociation (Maybe AssociationComplianceSeverity)

The severity level to assign to the association.

createAssociation_automationTargetParameterName :: Lens' CreateAssociation (Maybe Text)

Specify the target for the association. This target is required for associations that use an Automation runbook and target resources by using rate controls. Automation is a capability of Amazon Web Services Systems Manager.

createAssociation_targets :: Lens' CreateAssociation (Maybe [Target])

The targets for the association. You can target instances by using tags, Amazon Web Services resource groups, all instances in an Amazon Web Services account, or individual instance IDs. For more information about choosing targets for an association, see Using targets and rate controls with State Manager associations in the Amazon Web Services Systems Manager User Guide.

createAssociation_scheduleExpression :: Lens' CreateAssociation (Maybe Text)

A cron expression when the association will be applied to the target(s).

createAssociation_targetLocations :: Lens' CreateAssociation (Maybe (NonEmpty TargetLocation))

A location is a combination of Amazon Web Services Regions and Amazon Web Services accounts where you want to run the association. Use this action to create an association in multiple Regions and multiple accounts.

createAssociation_maxConcurrency :: Lens' CreateAssociation (Maybe Text)

The maximum number of targets allowed to run the association at the same time. You can specify a number, for example 10, or a percentage of the target set, for example 10%. The default value is 100%, which means all targets run the association at the same time.

If a new instance starts and attempts to run an association while Systems Manager is running MaxConcurrency associations, the association is allowed to run. During the next association interval, the new instance will process its association within the limit specified for MaxConcurrency.

createAssociation_calendarNames :: Lens' CreateAssociation (Maybe [Text])

The names or Amazon Resource Names (ARNs) of the Change Calendar type documents you want to gate your associations under. The associations only run when that change calendar is open. For more information, see Amazon Web Services Systems Manager Change Calendar.

createAssociation_associationName :: Lens' CreateAssociation (Maybe Text)

Specify a descriptive name for the association.

createAssociation_documentVersion :: Lens' CreateAssociation (Maybe Text)

The document version you want to associate with the target(s). Can be a specific version or the default version.

createAssociation_parameters :: Lens' CreateAssociation (Maybe (HashMap Text [Text]))

The parameters for the runtime configuration of the document.

createAssociation_syncCompliance :: Lens' CreateAssociation (Maybe AssociationSyncCompliance)

The mode for generating association compliance. You can specify AUTO or MANUAL. In AUTO mode, the system uses the status of the association execution to determine the compliance status. If the association execution runs successfully, then the association is COMPLIANT. If the association execution doesn't run successfully, the association is NON-COMPLIANT.

In MANUAL mode, you must specify the AssociationId as a parameter for the PutComplianceItems API operation. In this case, compliance data isn't managed by State Manager. It is managed by your direct call to the PutComplianceItems API operation.

By default, all associations use AUTO mode.

createAssociation_outputLocation :: Lens' CreateAssociation (Maybe InstanceAssociationOutputLocation)

An Amazon Simple Storage Service (Amazon S3) bucket where you want to store the output details of the request.

createAssociation_applyOnlyAtCronInterval :: Lens' CreateAssociation (Maybe Bool)

By default, when you create a new association, the system runs it immediately after it is created and then according to the schedule you specified. Specify this option if you don't want an association to run immediately after you create it. This parameter isn't supported for rate expressions.

createAssociation_name :: Lens' CreateAssociation Text

The name of the SSM Command document or Automation runbook that contains the configuration information for the instance.

You can specify Amazon Web Services-predefined documents, documents you created, or a document that is shared with you from another account.

For Systems Manager documents (SSM documents) that are shared with you from other Amazon Web Services accounts, you must specify the complete SSM document ARN, in the following format:

arn:partition:ssm:region:account-id:document/document-name

For example:

arn:aws:ssm:us-east-2:12345678912:document/My-Shared-Document

For Amazon Web Services-predefined documents and SSM documents you created in your account, you only need to specify the document name. For example, AWS-ApplyPatchBaseline or My-Document.

Destructuring the Response

data CreateAssociationResponse

See: newCreateAssociationResponse smart constructor.

Constructors

CreateAssociationResponse'

Fields

associationDescription :: Maybe AssociationDescription

Information about the association.

httpStatus :: Int

The response's http status code.

newCreateAssociationResponse

Arguments

:: Int
$sel:httpStatus:CreateAssociationResponse'

-> CreateAssociationResponse

Create a value of CreateAssociationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associationDescription:CreateAssociationResponse', createAssociationResponse_associationDescription - Information about the association.

$sel:httpStatus:CreateAssociationResponse', createAssociationResponse_httpStatus - The response's http status code.

Response Lenses

createAssociationResponse_associationDescription :: Lens' CreateAssociationResponse (Maybe AssociationDescription)

Information about the association.

createAssociationResponse_httpStatus :: Lens' CreateAssociationResponse Int

The response's http status code.