Network.AWS.SSM.CreateOpsItem

Description

Creates a new OpsItem. You must have permission in Identity and Access Management (IAM) to create a new OpsItem. For more information, see Getting started with OpsCenter in the Amazon Web Services Systems Manager User Guide.

Operations engineers and IT professionals use Amazon Web Services Systems Manager OpsCenter to view, investigate, and remediate operational issues impacting the performance and health of their Amazon Web Services resources. For more information, see Amazon Web Services Systems Manager OpsCenter in the Amazon Web Services Systems Manager User Guide.

Creating a Request

data CreateOpsItem

See: newCreateOpsItem smart constructor.

Constructors

CreateOpsItem'

Fields

plannedEndTime :: Maybe POSIX

The time specified in a change request for a runbook workflow to end. Currently supported only for the OpsItem type /aws/changerequest.

severity :: Maybe Text

Specify a severity to assign to an OpsItem.

actualStartTime :: Maybe POSIX

The time a runbook workflow started. Currently reported only for the OpsItem type /aws/changerequest.

category :: Maybe Text

Specify a category to assign to an OpsItem.

operationalData :: Maybe (HashMap Text OpsItemDataValue)

Operational data is custom data that provides useful reference details about the OpsItem. For example, you can specify log files, error strings, license keys, troubleshooting tips, or other relevant data. You enter operational data as key-value pairs. The key has a maximum length of 128 characters. The value has a maximum size of 20 KB.

Operational data keys can't begin with the following: amazon, aws, amzn, ssm, /amazon, /aws, /amzn, /ssm.

You can choose to make the data searchable by other users in the account or you can restrict search access. Searchable data means that all users with access to the OpsItem Overview page (as provided by the DescribeOpsItems API operation) can view and search on the specified data. Operational data that isn't searchable is only viewable by users who have access to the OpsItem (as provided by the GetOpsItem API operation).

Use the /aws/resources key in OperationalData to specify a related resource in the request. Use the /aws/automations key in OperationalData to associate an Automation runbook with the OpsItem. To view Amazon Web Services CLI example commands that use these keys, see Creating OpsItems manually in the Amazon Web Services Systems Manager User Guide.

priority :: Maybe Natural

The importance of this OpsItem in relation to other OpsItems in the system.

actualEndTime :: Maybe POSIX

The time a runbook workflow ended. Currently reported only for the OpsItem type /aws/changerequest.

tags :: Maybe [Tag]

Optional metadata that you assign to a resource. You can restrict access to OpsItems by using an inline IAM policy that specifies tags. For more information, see Getting started with OpsCenter in the Amazon Web Services Systems Manager User Guide.

Tags use a key-value pair. For example:

Key=Department,Value=Finance

To add tags to a new OpsItem, a user must have IAM permissions for both the ssm:CreateOpsItems operation and the ssm:AddTagsToResource operation. To add tags to an existing OpsItem, use the AddTagsToResource operation.

plannedStartTime :: Maybe POSIX

The time specified in a change request for a runbook workflow to start. Currently supported only for the OpsItem type /aws/changerequest.

opsItemType :: Maybe Text

The type of OpsItem to create. Currently, the only valid values are /aws/changerequest and /aws/issue.

notifications :: Maybe [OpsItemNotification]

The Amazon Resource Name (ARN) of an SNS topic where notifications are sent when this OpsItem is edited or changed.

relatedOpsItems :: Maybe [RelatedOpsItem]

One or more OpsItems that share something in common with the current OpsItems. For example, related OpsItems can include OpsItems with similar error messages, impacted resources, or statuses for the impacted resource.

description :: Text

Information about the OpsItem.

source :: Text

The origin of the OpsItem, such as Amazon EC2 or Systems Manager.

The source name can't contain the following strings: aws, amazon, and amzn.

title :: Text

A short heading that describes the nature of the OpsItem and the impacted resource.

newCreateOpsItem

Arguments

:: Text
$sel:description:CreateOpsItem'

-> Text
$sel:source:CreateOpsItem'

-> Text
$sel:title:CreateOpsItem'

-> CreateOpsItem

Create a value of CreateOpsItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:plannedEndTime:CreateOpsItem', createOpsItem_plannedEndTime - The time specified in a change request for a runbook workflow to end. Currently supported only for the OpsItem type /aws/changerequest.

$sel:severity:CreateOpsItem', createOpsItem_severity - Specify a severity to assign to an OpsItem.

$sel:actualStartTime:CreateOpsItem', createOpsItem_actualStartTime - The time a runbook workflow started. Currently reported only for the OpsItem type /aws/changerequest.

$sel:category:CreateOpsItem', createOpsItem_category - Specify a category to assign to an OpsItem.

$sel:operationalData:CreateOpsItem', createOpsItem_operationalData - Operational data is custom data that provides useful reference details about the OpsItem. For example, you can specify log files, error strings, license keys, troubleshooting tips, or other relevant data. You enter operational data as key-value pairs. The key has a maximum length of 128 characters. The value has a maximum size of 20 KB.

Operational data keys can't begin with the following: amazon, aws, amzn, ssm, /amazon, /aws, /amzn, /ssm.

You can choose to make the data searchable by other users in the account or you can restrict search access. Searchable data means that all users with access to the OpsItem Overview page (as provided by the DescribeOpsItems API operation) can view and search on the specified data. Operational data that isn't searchable is only viewable by users who have access to the OpsItem (as provided by the GetOpsItem API operation).

Use the /aws/resources key in OperationalData to specify a related resource in the request. Use the /aws/automations key in OperationalData to associate an Automation runbook with the OpsItem. To view Amazon Web Services CLI example commands that use these keys, see Creating OpsItems manually in the Amazon Web Services Systems Manager User Guide.

$sel:priority:CreateOpsItem', createOpsItem_priority - The importance of this OpsItem in relation to other OpsItems in the system.

$sel:actualEndTime:CreateOpsItem', createOpsItem_actualEndTime - The time a runbook workflow ended. Currently reported only for the OpsItem type /aws/changerequest.

$sel:tags:CreateOpsItem', createOpsItem_tags - Optional metadata that you assign to a resource. You can restrict access to OpsItems by using an inline IAM policy that specifies tags. For more information, see Getting started with OpsCenter in the Amazon Web Services Systems Manager User Guide.

Tags use a key-value pair. For example:

Key=Department,Value=Finance

To add tags to a new OpsItem, a user must have IAM permissions for both the ssm:CreateOpsItems operation and the ssm:AddTagsToResource operation. To add tags to an existing OpsItem, use the AddTagsToResource operation.

$sel:plannedStartTime:CreateOpsItem', createOpsItem_plannedStartTime - The time specified in a change request for a runbook workflow to start. Currently supported only for the OpsItem type /aws/changerequest.

$sel:opsItemType:CreateOpsItem', createOpsItem_opsItemType - The type of OpsItem to create. Currently, the only valid values are /aws/changerequest and /aws/issue.

$sel:notifications:CreateOpsItem', createOpsItem_notifications - The Amazon Resource Name (ARN) of an SNS topic where notifications are sent when this OpsItem is edited or changed.

$sel:relatedOpsItems:CreateOpsItem', createOpsItem_relatedOpsItems - One or more OpsItems that share something in common with the current OpsItems. For example, related OpsItems can include OpsItems with similar error messages, impacted resources, or statuses for the impacted resource.

$sel:description:CreateOpsItem', createOpsItem_description - Information about the OpsItem.

$sel:source:CreateOpsItem', createOpsItem_source - The origin of the OpsItem, such as Amazon EC2 or Systems Manager.

The source name can't contain the following strings: aws, amazon, and amzn.

$sel:title:CreateOpsItem', createOpsItem_title - A short heading that describes the nature of the OpsItem and the impacted resource.

Request Lenses

createOpsItem_plannedEndTime :: Lens' CreateOpsItem (Maybe UTCTime)

The time specified in a change request for a runbook workflow to end. Currently supported only for the OpsItem type /aws/changerequest.

createOpsItem_severity :: Lens' CreateOpsItem (Maybe Text)

Specify a severity to assign to an OpsItem.

createOpsItem_actualStartTime :: Lens' CreateOpsItem (Maybe UTCTime)

The time a runbook workflow started. Currently reported only for the OpsItem type /aws/changerequest.

createOpsItem_category :: Lens' CreateOpsItem (Maybe Text)

Specify a category to assign to an OpsItem.

createOpsItem_operationalData :: Lens' CreateOpsItem (Maybe (HashMap Text OpsItemDataValue))

Operational data is custom data that provides useful reference details about the OpsItem. For example, you can specify log files, error strings, license keys, troubleshooting tips, or other relevant data. You enter operational data as key-value pairs. The key has a maximum length of 128 characters. The value has a maximum size of 20 KB.

Operational data keys can't begin with the following: amazon, aws, amzn, ssm, /amazon, /aws, /amzn, /ssm.

You can choose to make the data searchable by other users in the account or you can restrict search access. Searchable data means that all users with access to the OpsItem Overview page (as provided by the DescribeOpsItems API operation) can view and search on the specified data. Operational data that isn't searchable is only viewable by users who have access to the OpsItem (as provided by the GetOpsItem API operation).

Use the /aws/resources key in OperationalData to specify a related resource in the request. Use the /aws/automations key in OperationalData to associate an Automation runbook with the OpsItem. To view Amazon Web Services CLI example commands that use these keys, see Creating OpsItems manually in the Amazon Web Services Systems Manager User Guide.

createOpsItem_priority :: Lens' CreateOpsItem (Maybe Natural)

The importance of this OpsItem in relation to other OpsItems in the system.

createOpsItem_actualEndTime :: Lens' CreateOpsItem (Maybe UTCTime)

The time a runbook workflow ended. Currently reported only for the OpsItem type /aws/changerequest.

createOpsItem_tags :: Lens' CreateOpsItem (Maybe [Tag])

Optional metadata that you assign to a resource. You can restrict access to OpsItems by using an inline IAM policy that specifies tags. For more information, see Getting started with OpsCenter in the Amazon Web Services Systems Manager User Guide.

Tags use a key-value pair. For example:

Key=Department,Value=Finance

To add tags to a new OpsItem, a user must have IAM permissions for both the ssm:CreateOpsItems operation and the ssm:AddTagsToResource operation. To add tags to an existing OpsItem, use the AddTagsToResource operation.

createOpsItem_plannedStartTime :: Lens' CreateOpsItem (Maybe UTCTime)

The time specified in a change request for a runbook workflow to start. Currently supported only for the OpsItem type /aws/changerequest.

createOpsItem_opsItemType :: Lens' CreateOpsItem (Maybe Text)

The type of OpsItem to create. Currently, the only valid values are /aws/changerequest and /aws/issue.

createOpsItem_notifications :: Lens' CreateOpsItem (Maybe [OpsItemNotification])

The Amazon Resource Name (ARN) of an SNS topic where notifications are sent when this OpsItem is edited or changed.

createOpsItem_relatedOpsItems :: Lens' CreateOpsItem (Maybe [RelatedOpsItem])

One or more OpsItems that share something in common with the current OpsItems. For example, related OpsItems can include OpsItems with similar error messages, impacted resources, or statuses for the impacted resource.

createOpsItem_description :: Lens' CreateOpsItem Text

Information about the OpsItem.

createOpsItem_source :: Lens' CreateOpsItem Text

The origin of the OpsItem, such as Amazon EC2 or Systems Manager.

The source name can't contain the following strings: aws, amazon, and amzn.

createOpsItem_title :: Lens' CreateOpsItem Text

A short heading that describes the nature of the OpsItem and the impacted resource.

Destructuring the Response

data CreateOpsItemResponse

See: newCreateOpsItemResponse smart constructor.

Constructors

CreateOpsItemResponse'

Fields

opsItemId :: Maybe Text

The ID of the OpsItem.

httpStatus :: Int

The response's http status code.

newCreateOpsItemResponse

Arguments

:: Int
$sel:httpStatus:CreateOpsItemResponse'

-> CreateOpsItemResponse

Create a value of CreateOpsItemResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:opsItemId:CreateOpsItemResponse', createOpsItemResponse_opsItemId - The ID of the OpsItem.

$sel:httpStatus:CreateOpsItemResponse', createOpsItemResponse_httpStatus - The response's http status code.

Response Lenses

createOpsItemResponse_opsItemId :: Lens' CreateOpsItemResponse (Maybe Text)

The ID of the OpsItem.

createOpsItemResponse_httpStatus :: Lens' CreateOpsItemResponse Int

The response's http status code.