Network.AWS.SSM.CreateOpsMetadata

Description

If you create a new application in Application Manager, Amazon Web Services Systems Manager calls this API operation to specify information about the new application, including the application type.

Creating a Request

data CreateOpsMetadata

See: newCreateOpsMetadata smart constructor.

Constructors

CreateOpsMetadata'

Fields

metadata :: Maybe (HashMap Text MetadataValue)

Metadata for a new Application Manager application.

tags :: Maybe [Tag]

Optional metadata that you assign to a resource. You can specify a maximum of five tags for an OpsMetadata object. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag an OpsMetadata object to identify an environment or target Amazon Web Services Region. In this case, you could specify the following key-value pairs:

Key=Environment,Value=Production

Key=Region,Value=us-east-2

resourceId :: Text

A resource ID for a new Application Manager application.

newCreateOpsMetadata

Arguments

:: Text
$sel:resourceId:CreateOpsMetadata'

-> CreateOpsMetadata

Create a value of CreateOpsMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metadata:CreateOpsMetadata', createOpsMetadata_metadata - Metadata for a new Application Manager application.

$sel:tags:CreateOpsMetadata', createOpsMetadata_tags - Optional metadata that you assign to a resource. You can specify a maximum of five tags for an OpsMetadata object. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag an OpsMetadata object to identify an environment or target Amazon Web Services Region. In this case, you could specify the following key-value pairs:

Key=Environment,Value=Production

Key=Region,Value=us-east-2

$sel:resourceId:CreateOpsMetadata', createOpsMetadata_resourceId - A resource ID for a new Application Manager application.

Request Lenses

createOpsMetadata_metadata :: Lens' CreateOpsMetadata (Maybe (HashMap Text MetadataValue))

Metadata for a new Application Manager application.

createOpsMetadata_tags :: Lens' CreateOpsMetadata (Maybe [Tag])

Optional metadata that you assign to a resource. You can specify a maximum of five tags for an OpsMetadata object. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag an OpsMetadata object to identify an environment or target Amazon Web Services Region. In this case, you could specify the following key-value pairs:

Key=Environment,Value=Production

Key=Region,Value=us-east-2

createOpsMetadata_resourceId :: Lens' CreateOpsMetadata Text

A resource ID for a new Application Manager application.

Destructuring the Response

data CreateOpsMetadataResponse

See: newCreateOpsMetadataResponse smart constructor.

Constructors

CreateOpsMetadataResponse'

Fields

opsMetadataArn :: Maybe Text

The Amazon Resource Name (ARN) of the OpsMetadata Object or blob created by the call.

httpStatus :: Int

The response's http status code.

newCreateOpsMetadataResponse

Arguments

:: Int
$sel:httpStatus:CreateOpsMetadataResponse'

-> CreateOpsMetadataResponse

Create a value of CreateOpsMetadataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:opsMetadataArn:CreateOpsMetadataResponse', createOpsMetadataResponse_opsMetadataArn - The Amazon Resource Name (ARN) of the OpsMetadata Object or blob created by the call.

$sel:httpStatus:CreateOpsMetadataResponse', createOpsMetadataResponse_httpStatus - The response's http status code.

Response Lenses

createOpsMetadataResponse_opsMetadataArn :: Lens' CreateOpsMetadataResponse (Maybe Text)

The Amazon Resource Name (ARN) of the OpsMetadata Object or blob created by the call.

createOpsMetadataResponse_httpStatus :: Lens' CreateOpsMetadataResponse Int

The response's http status code.