Network.AWS.SSM.CreatePatchBaseline

Description

Creates a patch baseline.

For information about valid key-value pairs in PatchFilters for each supported operating system type, see PatchFilter.

Creating a Request

data CreatePatchBaseline

See: newCreatePatchBaseline smart constructor.

Constructors

CreatePatchBaseline'

Fields

approvalRules :: Maybe PatchRuleGroup

A set of rules used to include patches in the baseline.

clientToken :: Maybe Text

User-provided idempotency token.

operatingSystem :: Maybe OperatingSystem

Defines the operating system the patch baseline applies to. The default value is WINDOWS.

globalFilters :: Maybe PatchFilterGroup

A set of global filters used to include patches in the baseline.

approvedPatchesComplianceLevel :: Maybe PatchComplianceLevel

Defines the compliance level for approved patches. When an approved patch is reported as missing, this value describes the severity of the compliance violation. The default value is UNSPECIFIED.

rejectedPatchesAction :: Maybe PatchAction

The action for Patch Manager to take on patches included in the RejectedPackages list.

ALLOW_AS_DEPENDENCY : A package in the Rejected patches list is installed only if it is a dependency of another package. It is considered compliant with the patch baseline, and its status is reported as InstalledOther. This is the default action if no option is specified.

BLOCK : Packages in the RejectedPatches list, and packages that include them as dependencies, aren't installed under any circumstances. If a package was installed before it was added to the Rejected patches list, it is considered non-compliant with the patch baseline, and its status is reported as InstalledRejected.

approvedPatches :: Maybe [Text]

A list of explicitly approved patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

approvedPatchesEnableNonSecurity :: Maybe Bool

Indicates whether the list of approved patches includes non-security updates that should be applied to the instances. The default value is false. Applies to Linux instances only.

rejectedPatches :: Maybe [Text]

A list of explicitly rejected patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

sources :: Maybe [PatchSource]

Information about the patches to use to update the instances, including target operating systems and source repositories. Applies to Linux instances only.

description :: Maybe Text

A description of the patch baseline.

tags :: Maybe [Tag]

Optional metadata that you assign to a resource. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag a patch baseline to identify the severity level of patches it specifies and the operating system family it applies to. In this case, you could specify the following key-value pairs:

Key=PatchSeverity,Value=Critical

Key=OS,Value=Windows

To add tags to an existing patch baseline, use the AddTagsToResource operation.

name :: Text

The name of the patch baseline.

newCreatePatchBaseline

Arguments

:: Text
$sel:name:CreatePatchBaseline'

-> CreatePatchBaseline

Create a value of CreatePatchBaseline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:approvalRules:CreatePatchBaseline', createPatchBaseline_approvalRules - A set of rules used to include patches in the baseline.

$sel:clientToken:CreatePatchBaseline', createPatchBaseline_clientToken - User-provided idempotency token.

$sel:operatingSystem:CreatePatchBaseline', createPatchBaseline_operatingSystem - Defines the operating system the patch baseline applies to. The default value is WINDOWS.

$sel:globalFilters:CreatePatchBaseline', createPatchBaseline_globalFilters - A set of global filters used to include patches in the baseline.

$sel:approvedPatchesComplianceLevel:CreatePatchBaseline', createPatchBaseline_approvedPatchesComplianceLevel - Defines the compliance level for approved patches. When an approved patch is reported as missing, this value describes the severity of the compliance violation. The default value is UNSPECIFIED.

$sel:rejectedPatchesAction:CreatePatchBaseline', createPatchBaseline_rejectedPatchesAction - The action for Patch Manager to take on patches included in the RejectedPackages list.

ALLOW_AS_DEPENDENCY : A package in the Rejected patches list is installed only if it is a dependency of another package. It is considered compliant with the patch baseline, and its status is reported as InstalledOther. This is the default action if no option is specified.

BLOCK : Packages in the RejectedPatches list, and packages that include them as dependencies, aren't installed under any circumstances. If a package was installed before it was added to the Rejected patches list, it is considered non-compliant with the patch baseline, and its status is reported as InstalledRejected.

$sel:approvedPatches:CreatePatchBaseline', createPatchBaseline_approvedPatches - A list of explicitly approved patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

$sel:approvedPatchesEnableNonSecurity:CreatePatchBaseline', createPatchBaseline_approvedPatchesEnableNonSecurity - Indicates whether the list of approved patches includes non-security updates that should be applied to the instances. The default value is false. Applies to Linux instances only.

$sel:rejectedPatches:CreatePatchBaseline', createPatchBaseline_rejectedPatches - A list of explicitly rejected patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

$sel:sources:CreatePatchBaseline', createPatchBaseline_sources - Information about the patches to use to update the instances, including target operating systems and source repositories. Applies to Linux instances only.

$sel:description:CreatePatchBaseline', createPatchBaseline_description - A description of the patch baseline.

$sel:tags:CreatePatchBaseline', createPatchBaseline_tags - Optional metadata that you assign to a resource. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag a patch baseline to identify the severity level of patches it specifies and the operating system family it applies to. In this case, you could specify the following key-value pairs:

Key=PatchSeverity,Value=Critical

Key=OS,Value=Windows

To add tags to an existing patch baseline, use the AddTagsToResource operation.

$sel:name:CreatePatchBaseline', createPatchBaseline_name - The name of the patch baseline.

Request Lenses

createPatchBaseline_approvalRules :: Lens' CreatePatchBaseline (Maybe PatchRuleGroup)

A set of rules used to include patches in the baseline.

createPatchBaseline_clientToken :: Lens' CreatePatchBaseline (Maybe Text)

User-provided idempotency token.

createPatchBaseline_operatingSystem :: Lens' CreatePatchBaseline (Maybe OperatingSystem)

Defines the operating system the patch baseline applies to. The default value is WINDOWS.

createPatchBaseline_globalFilters :: Lens' CreatePatchBaseline (Maybe PatchFilterGroup)

A set of global filters used to include patches in the baseline.

createPatchBaseline_approvedPatchesComplianceLevel :: Lens' CreatePatchBaseline (Maybe PatchComplianceLevel)

Defines the compliance level for approved patches. When an approved patch is reported as missing, this value describes the severity of the compliance violation. The default value is UNSPECIFIED.

createPatchBaseline_rejectedPatchesAction :: Lens' CreatePatchBaseline (Maybe PatchAction)

The action for Patch Manager to take on patches included in the RejectedPackages list.

ALLOW_AS_DEPENDENCY : A package in the Rejected patches list is installed only if it is a dependency of another package. It is considered compliant with the patch baseline, and its status is reported as InstalledOther. This is the default action if no option is specified.

BLOCK : Packages in the RejectedPatches list, and packages that include them as dependencies, aren't installed under any circumstances. If a package was installed before it was added to the Rejected patches list, it is considered non-compliant with the patch baseline, and its status is reported as InstalledRejected.

createPatchBaseline_approvedPatches :: Lens' CreatePatchBaseline (Maybe [Text])

A list of explicitly approved patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

createPatchBaseline_approvedPatchesEnableNonSecurity :: Lens' CreatePatchBaseline (Maybe Bool)

Indicates whether the list of approved patches includes non-security updates that should be applied to the instances. The default value is false. Applies to Linux instances only.

createPatchBaseline_rejectedPatches :: Lens' CreatePatchBaseline (Maybe [Text])

A list of explicitly rejected patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

createPatchBaseline_sources :: Lens' CreatePatchBaseline (Maybe [PatchSource])

Information about the patches to use to update the instances, including target operating systems and source repositories. Applies to Linux instances only.

createPatchBaseline_description :: Lens' CreatePatchBaseline (Maybe Text)

A description of the patch baseline.

createPatchBaseline_tags :: Lens' CreatePatchBaseline (Maybe [Tag])

Optional metadata that you assign to a resource. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag a patch baseline to identify the severity level of patches it specifies and the operating system family it applies to. In this case, you could specify the following key-value pairs:

Key=PatchSeverity,Value=Critical

Key=OS,Value=Windows

To add tags to an existing patch baseline, use the AddTagsToResource operation.

createPatchBaseline_name :: Lens' CreatePatchBaseline Text

The name of the patch baseline.

Destructuring the Response

data CreatePatchBaselineResponse

See: newCreatePatchBaselineResponse smart constructor.

Constructors

CreatePatchBaselineResponse'

Fields

baselineId :: Maybe Text

The ID of the created patch baseline.

httpStatus :: Int

The response's http status code.

newCreatePatchBaselineResponse

Arguments

:: Int
$sel:httpStatus:CreatePatchBaselineResponse'

-> CreatePatchBaselineResponse

Create a value of CreatePatchBaselineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baselineId:CreatePatchBaselineResponse', createPatchBaselineResponse_baselineId - The ID of the created patch baseline.

$sel:httpStatus:CreatePatchBaselineResponse', createPatchBaselineResponse_httpStatus - The response's http status code.

Response Lenses

createPatchBaselineResponse_baselineId :: Lens' CreatePatchBaselineResponse (Maybe Text)

The ID of the created patch baseline.

createPatchBaselineResponse_httpStatus :: Lens' CreatePatchBaselineResponse Int

The response's http status code.