Network.AWS.SSM.DeleteInventory

Description

Delete a custom inventory type or the data associated with a custom Inventory type. Deleting a custom inventory type is also referred to as deleting a custom inventory schema.

Creating a Request

data DeleteInventory

See: newDeleteInventory smart constructor.

Constructors

DeleteInventory'

Fields

dryRun :: Maybe Bool

Use this option to view a summary of the deletion request without deleting any data or the data type. This option is useful when you only want to understand what will be deleted. Once you validate that the data to be deleted is what you intend to delete, you can run the same command without specifying the DryRun option.

schemaDeleteOption :: Maybe InventorySchemaDeleteOption

Use the SchemaDeleteOption to delete a custom inventory type (schema). If you don't choose this option, the system only deletes existing inventory data associated with the custom inventory type. Choose one of the following options:

DisableSchema: If you choose this option, the system ignores all inventory data for the specified version, and any earlier versions. To enable this schema again, you must call the PutInventory operation for a version greater than the disabled version.

DeleteSchema: This option deletes the specified custom type from the Inventory service. You can recreate the schema later, if you want.

clientToken :: Maybe Text

User-provided idempotency token.

typeName :: Text

The name of the custom inventory type for which you want to delete either all previously collected data or the inventory type itself.

newDeleteInventory

Arguments

:: Text
$sel:typeName:DeleteInventory'

-> DeleteInventory

Create a value of DeleteInventory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DeleteInventory', deleteInventory_dryRun - Use this option to view a summary of the deletion request without deleting any data or the data type. This option is useful when you only want to understand what will be deleted. Once you validate that the data to be deleted is what you intend to delete, you can run the same command without specifying the DryRun option.

$sel:schemaDeleteOption:DeleteInventory', deleteInventory_schemaDeleteOption - Use the SchemaDeleteOption to delete a custom inventory type (schema). If you don't choose this option, the system only deletes existing inventory data associated with the custom inventory type. Choose one of the following options:

DisableSchema: If you choose this option, the system ignores all inventory data for the specified version, and any earlier versions. To enable this schema again, you must call the PutInventory operation for a version greater than the disabled version.

DeleteSchema: This option deletes the specified custom type from the Inventory service. You can recreate the schema later, if you want.

$sel:clientToken:DeleteInventory', deleteInventory_clientToken - User-provided idempotency token.

$sel:typeName:DeleteInventory', deleteInventory_typeName - The name of the custom inventory type for which you want to delete either all previously collected data or the inventory type itself.

Request Lenses

deleteInventory_dryRun :: Lens' DeleteInventory (Maybe Bool)

Use this option to view a summary of the deletion request without deleting any data or the data type. This option is useful when you only want to understand what will be deleted. Once you validate that the data to be deleted is what you intend to delete, you can run the same command without specifying the DryRun option.

deleteInventory_schemaDeleteOption :: Lens' DeleteInventory (Maybe InventorySchemaDeleteOption)

Use the SchemaDeleteOption to delete a custom inventory type (schema). If you don't choose this option, the system only deletes existing inventory data associated with the custom inventory type. Choose one of the following options:

DisableSchema: If you choose this option, the system ignores all inventory data for the specified version, and any earlier versions. To enable this schema again, you must call the PutInventory operation for a version greater than the disabled version.

DeleteSchema: This option deletes the specified custom type from the Inventory service. You can recreate the schema later, if you want.

deleteInventory_clientToken :: Lens' DeleteInventory (Maybe Text)

User-provided idempotency token.

deleteInventory_typeName :: Lens' DeleteInventory Text

The name of the custom inventory type for which you want to delete either all previously collected data or the inventory type itself.

Destructuring the Response

data DeleteInventoryResponse

See: newDeleteInventoryResponse smart constructor.

Constructors

DeleteInventoryResponse'

Fields

typeName :: Maybe Text

The name of the inventory data type specified in the request.

deletionId :: Maybe Text

Every DeleteInventory operation is assigned a unique ID. This option returns a unique ID. You can use this ID to query the status of a delete operation. This option is useful for ensuring that a delete operation has completed before you begin other operations.

deletionSummary :: Maybe InventoryDeletionSummary

A summary of the delete operation. For more information about this summary, see Deleting custom inventory in the Amazon Web Services Systems Manager User Guide.

httpStatus :: Int

The response's http status code.

newDeleteInventoryResponse

Arguments

:: Int
$sel:httpStatus:DeleteInventoryResponse'

-> DeleteInventoryResponse

Create a value of DeleteInventoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeName:DeleteInventory', deleteInventoryResponse_typeName - The name of the inventory data type specified in the request.

$sel:deletionId:DeleteInventoryResponse', deleteInventoryResponse_deletionId - Every DeleteInventory operation is assigned a unique ID. This option returns a unique ID. You can use this ID to query the status of a delete operation. This option is useful for ensuring that a delete operation has completed before you begin other operations.

$sel:deletionSummary:DeleteInventoryResponse', deleteInventoryResponse_deletionSummary - A summary of the delete operation. For more information about this summary, see Deleting custom inventory in the Amazon Web Services Systems Manager User Guide.

$sel:httpStatus:DeleteInventoryResponse', deleteInventoryResponse_httpStatus - The response's http status code.

Response Lenses

deleteInventoryResponse_typeName :: Lens' DeleteInventoryResponse (Maybe Text)

The name of the inventory data type specified in the request.

deleteInventoryResponse_deletionId :: Lens' DeleteInventoryResponse (Maybe Text)

Every DeleteInventory operation is assigned a unique ID. This option returns a unique ID. You can use this ID to query the status of a delete operation. This option is useful for ensuring that a delete operation has completed before you begin other operations.

deleteInventoryResponse_deletionSummary :: Lens' DeleteInventoryResponse (Maybe InventoryDeletionSummary)

A summary of the delete operation. For more information about this summary, see Deleting custom inventory in the Amazon Web Services Systems Manager User Guide.

deleteInventoryResponse_httpStatus :: Lens' DeleteInventoryResponse Int

The response's http status code.