Network.AWS.SSM.DeleteParameters

Description

Delete a list of parameters. After deleting a parameter, wait for at least 30 seconds to create a parameter with the same name.

Creating a Request

data DeleteParameters

See: newDeleteParameters smart constructor.

Constructors

DeleteParameters'

Fields

names :: NonEmpty Text

The names of the parameters to delete. After deleting a parameter, wait for at least 30 seconds to create a parameter with the same name.

newDeleteParameters

Arguments

:: NonEmpty Text
$sel:names:DeleteParameters'

-> DeleteParameters

Create a value of DeleteParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:names:DeleteParameters', deleteParameters_names - The names of the parameters to delete. After deleting a parameter, wait for at least 30 seconds to create a parameter with the same name.

Request Lenses

deleteParameters_names :: Lens' DeleteParameters (NonEmpty Text)

The names of the parameters to delete. After deleting a parameter, wait for at least 30 seconds to create a parameter with the same name.

Destructuring the Response

data DeleteParametersResponse

See: newDeleteParametersResponse smart constructor.

Constructors

DeleteParametersResponse'

Fields

invalidParameters :: Maybe (NonEmpty Text)

The names of parameters that weren't deleted because the parameters aren't valid.

deletedParameters :: Maybe (NonEmpty Text)

The names of the deleted parameters.

httpStatus :: Int

The response's http status code.

newDeleteParametersResponse

Arguments

:: Int
$sel:httpStatus:DeleteParametersResponse'

-> DeleteParametersResponse

Create a value of DeleteParametersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:invalidParameters:DeleteParametersResponse', deleteParametersResponse_invalidParameters - The names of parameters that weren't deleted because the parameters aren't valid.

$sel:deletedParameters:DeleteParametersResponse', deleteParametersResponse_deletedParameters - The names of the deleted parameters.

$sel:httpStatus:DeleteParametersResponse', deleteParametersResponse_httpStatus - The response's http status code.

Response Lenses

deleteParametersResponse_invalidParameters :: Lens' DeleteParametersResponse (Maybe (NonEmpty Text))

The names of parameters that weren't deleted because the parameters aren't valid.

deleteParametersResponse_deletedParameters :: Lens' DeleteParametersResponse (Maybe (NonEmpty Text))

The names of the deleted parameters.

deleteParametersResponse_httpStatus :: Lens' DeleteParametersResponse Int

The response's http status code.