Network.AWS.SSM.DeregisterTargetFromMaintenanceWindow

Description

Removes a target from a maintenance window.

Creating a Request

data DeregisterTargetFromMaintenanceWindow

See: newDeregisterTargetFromMaintenanceWindow smart constructor.

Constructors

DeregisterTargetFromMaintenanceWindow'

Fields

safe :: Maybe Bool

The system checks if the target is being referenced by a task. If the target is being referenced, the system returns an error and doesn't deregister the target from the maintenance window.

windowId :: Text

The ID of the maintenance window the target should be removed from.

windowTargetId :: Text

The ID of the target definition to remove.

newDeregisterTargetFromMaintenanceWindow

Arguments

:: Text
$sel:windowId:DeregisterTargetFromMaintenanceWindow'

-> Text
$sel:windowTargetId:DeregisterTargetFromMaintenanceWindow'

-> DeregisterTargetFromMaintenanceWindow

Create a value of DeregisterTargetFromMaintenanceWindow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:safe:DeregisterTargetFromMaintenanceWindow', deregisterTargetFromMaintenanceWindow_safe - The system checks if the target is being referenced by a task. If the target is being referenced, the system returns an error and doesn't deregister the target from the maintenance window.

$sel:windowId:DeregisterTargetFromMaintenanceWindow', deregisterTargetFromMaintenanceWindow_windowId - The ID of the maintenance window the target should be removed from.

$sel:windowTargetId:DeregisterTargetFromMaintenanceWindow', deregisterTargetFromMaintenanceWindow_windowTargetId - The ID of the target definition to remove.

Request Lenses

deregisterTargetFromMaintenanceWindow_safe :: Lens' DeregisterTargetFromMaintenanceWindow (Maybe Bool)

The system checks if the target is being referenced by a task. If the target is being referenced, the system returns an error and doesn't deregister the target from the maintenance window.

deregisterTargetFromMaintenanceWindow_windowId :: Lens' DeregisterTargetFromMaintenanceWindow Text

The ID of the maintenance window the target should be removed from.

deregisterTargetFromMaintenanceWindow_windowTargetId :: Lens' DeregisterTargetFromMaintenanceWindow Text

The ID of the target definition to remove.

Destructuring the Response

data DeregisterTargetFromMaintenanceWindowResponse

See: newDeregisterTargetFromMaintenanceWindowResponse smart constructor.

Constructors

DeregisterTargetFromMaintenanceWindowResponse'

Fields

windowTargetId :: Maybe Text

The ID of the removed target definition.

windowId :: Maybe Text

The ID of the maintenance window the target was removed from.

httpStatus :: Int

The response's http status code.

newDeregisterTargetFromMaintenanceWindowResponse

Arguments

:: Int
$sel:httpStatus:DeregisterTargetFromMaintenanceWindowResponse'

-> DeregisterTargetFromMaintenanceWindowResponse

Create a value of DeregisterTargetFromMaintenanceWindowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:windowTargetId:DeregisterTargetFromMaintenanceWindow', deregisterTargetFromMaintenanceWindowResponse_windowTargetId - The ID of the removed target definition.

$sel:windowId:DeregisterTargetFromMaintenanceWindow', deregisterTargetFromMaintenanceWindowResponse_windowId - The ID of the maintenance window the target was removed from.

$sel:httpStatus:DeregisterTargetFromMaintenanceWindowResponse', deregisterTargetFromMaintenanceWindowResponse_httpStatus - The response's http status code.

Response Lenses

deregisterTargetFromMaintenanceWindowResponse_windowTargetId :: Lens' DeregisterTargetFromMaintenanceWindowResponse (Maybe Text)

The ID of the removed target definition.

deregisterTargetFromMaintenanceWindowResponse_windowId :: Lens' DeregisterTargetFromMaintenanceWindowResponse (Maybe Text)

The ID of the maintenance window the target was removed from.

deregisterTargetFromMaintenanceWindowResponse_httpStatus :: Lens' DeregisterTargetFromMaintenanceWindowResponse Int

The response's http status code.