Network.AWS.SSM.DescribeActivations

Description

Describes details about the activation, such as the date and time the activation was created, its expiration date, the Identity and Access Management (IAM) role assigned to the instances in the activation, and the number of instances registered by using this activation.

This operation returns paginated results.

Creating a Request

data DescribeActivations

See: newDescribeActivations smart constructor.

Constructors

DescribeActivations'

Fields

filters :: Maybe [DescribeActivationsFilter]

A filter to view information about your activations.

nextToken :: Maybe Text

A token to start the list. Use this token to get the next set of results.

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

newDescribeActivations :: DescribeActivations

Create a value of DescribeActivations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeActivations', describeActivations_filters - A filter to view information about your activations.

$sel:nextToken:DescribeActivations', describeActivations_nextToken - A token to start the list. Use this token to get the next set of results.

$sel:maxResults:DescribeActivations', describeActivations_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

Request Lenses

describeActivations_filters :: Lens' DescribeActivations (Maybe [DescribeActivationsFilter])

A filter to view information about your activations.

describeActivations_nextToken :: Lens' DescribeActivations (Maybe Text)

A token to start the list. Use this token to get the next set of results.

describeActivations_maxResults :: Lens' DescribeActivations (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

Destructuring the Response

data DescribeActivationsResponse

See: newDescribeActivationsResponse smart constructor.

Constructors

DescribeActivationsResponse'

Fields

activationList :: Maybe [Activation]

A list of activations for your Amazon Web Services account.

nextToken :: Maybe Text

The token for the next set of items to return. Use this token to get the next set of results.

httpStatus :: Int

The response's http status code.

newDescribeActivationsResponse

Arguments

:: Int
$sel:httpStatus:DescribeActivationsResponse'

-> DescribeActivationsResponse

Create a value of DescribeActivationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:activationList:DescribeActivationsResponse', describeActivationsResponse_activationList - A list of activations for your Amazon Web Services account.

$sel:nextToken:DescribeActivations', describeActivationsResponse_nextToken - The token for the next set of items to return. Use this token to get the next set of results.

$sel:httpStatus:DescribeActivationsResponse', describeActivationsResponse_httpStatus - The response's http status code.

Response Lenses

describeActivationsResponse_activationList :: Lens' DescribeActivationsResponse (Maybe [Activation])

A list of activations for your Amazon Web Services account.

describeActivationsResponse_nextToken :: Lens' DescribeActivationsResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

describeActivationsResponse_httpStatus :: Lens' DescribeActivationsResponse Int

The response's http status code.