Network.AWS.SSM.DescribeAvailablePatches

Description

Lists all patches eligible to be included in a patch baseline.

This operation returns paginated results.

Creating a Request

data DescribeAvailablePatches

See: newDescribeAvailablePatches smart constructor.

Constructors

DescribeAvailablePatches'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

maxResults :: Maybe Natural

The maximum number of patches to return (per page).

filters :: Maybe [PatchOrchestratorFilter]

Each element in the array is a structure containing a key-value pair.

Windows Server

Supported keys for Windows Server instance patches include the following:

PATCH_SET

Sample values: OS | APPLICATION

PRODUCT

Sample values: WindowsServer2012 | Office 2010 | MicrosoftDefenderAntivirus

PRODUCT_FAMILY

Sample values: Windows | Office

MSRC_SEVERITY

Sample values: ServicePacks | Important | Moderate

CLASSIFICATION

Sample values: ServicePacks | SecurityUpdates | DefinitionUpdates

PATCH_ID

Sample values: KB123456 | KB4516046

Linux

When specifying filters for Linux patches, you must specify a key-pair for PRODUCT. For example, using the Command Line Interface (CLI), the following command fails:

aws ssm describe-available-patches --filters Key=CVE_ID,Values=CVE-2018-3615

However, the following command succeeds:

aws ssm describe-available-patches --filters Key=PRODUCT,Values=AmazonLinux2018.03 Key=CVE_ID,Values=CVE-2018-3615

Supported keys for Linux instance patches include the following:

PRODUCT

Sample values: AmazonLinux2018.03 | AmazonLinux2.0

NAME

Sample values: kernel-headers | samba-python | php

SEVERITY

Sample values: Critical | Important | Medium | Low

EPOCH

Sample values: 0 | 1

VERSION

Sample values: 78.6.1 | 4.10.16

RELEASE

Sample values: 9.56.amzn1 | 1.amzn2

ARCH

Sample values: i686 | x86_64

REPOSITORY

Sample values: Core | Updates

ADVISORY_ID

Sample values: ALAS-2018-1058 | ALAS2-2021-1594

CVE_ID

Sample values: CVE-2018-3615 | CVE-2020-1472

BUGZILLA_ID

Sample values: 1463241

newDescribeAvailablePatches :: DescribeAvailablePatches

Create a value of DescribeAvailablePatches with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeAvailablePatches', describeAvailablePatches_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:maxResults:DescribeAvailablePatches', describeAvailablePatches_maxResults - The maximum number of patches to return (per page).

$sel:filters:DescribeAvailablePatches', describeAvailablePatches_filters - Each element in the array is a structure containing a key-value pair.

Windows Server

Supported keys for Windows Server instance patches include the following:

PATCH_SET

Sample values: OS | APPLICATION

PRODUCT

Sample values: WindowsServer2012 | Office 2010 | MicrosoftDefenderAntivirus

PRODUCT_FAMILY

Sample values: Windows | Office

MSRC_SEVERITY

Sample values: ServicePacks | Important | Moderate

CLASSIFICATION

Sample values: ServicePacks | SecurityUpdates | DefinitionUpdates

PATCH_ID

Sample values: KB123456 | KB4516046

Linux

When specifying filters for Linux patches, you must specify a key-pair for PRODUCT. For example, using the Command Line Interface (CLI), the following command fails:

aws ssm describe-available-patches --filters Key=CVE_ID,Values=CVE-2018-3615

However, the following command succeeds:

aws ssm describe-available-patches --filters Key=PRODUCT,Values=AmazonLinux2018.03 Key=CVE_ID,Values=CVE-2018-3615

Supported keys for Linux instance patches include the following:

PRODUCT

Sample values: AmazonLinux2018.03 | AmazonLinux2.0

NAME

Sample values: kernel-headers | samba-python | php

SEVERITY

Sample values: Critical | Important | Medium | Low

EPOCH

Sample values: 0 | 1

VERSION

Sample values: 78.6.1 | 4.10.16

RELEASE

Sample values: 9.56.amzn1 | 1.amzn2

ARCH

Sample values: i686 | x86_64

REPOSITORY

Sample values: Core | Updates

ADVISORY_ID

Sample values: ALAS-2018-1058 | ALAS2-2021-1594

CVE_ID

Sample values: CVE-2018-3615 | CVE-2020-1472

BUGZILLA_ID

Sample values: 1463241

Request Lenses

describeAvailablePatches_nextToken :: Lens' DescribeAvailablePatches (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeAvailablePatches_maxResults :: Lens' DescribeAvailablePatches (Maybe Natural)

The maximum number of patches to return (per page).

describeAvailablePatches_filters :: Lens' DescribeAvailablePatches (Maybe [PatchOrchestratorFilter])

Each element in the array is a structure containing a key-value pair.

Windows Server

Supported keys for Windows Server instance patches include the following:

PATCH_SET

Sample values: OS | APPLICATION

PRODUCT

Sample values: WindowsServer2012 | Office 2010 | MicrosoftDefenderAntivirus

PRODUCT_FAMILY

Sample values: Windows | Office

MSRC_SEVERITY

Sample values: ServicePacks | Important | Moderate

CLASSIFICATION

Sample values: ServicePacks | SecurityUpdates | DefinitionUpdates

PATCH_ID

Sample values: KB123456 | KB4516046

Linux

When specifying filters for Linux patches, you must specify a key-pair for PRODUCT. For example, using the Command Line Interface (CLI), the following command fails:

aws ssm describe-available-patches --filters Key=CVE_ID,Values=CVE-2018-3615

However, the following command succeeds:

aws ssm describe-available-patches --filters Key=PRODUCT,Values=AmazonLinux2018.03 Key=CVE_ID,Values=CVE-2018-3615

Supported keys for Linux instance patches include the following:

PRODUCT

Sample values: AmazonLinux2018.03 | AmazonLinux2.0

NAME

Sample values: kernel-headers | samba-python | php

SEVERITY

Sample values: Critical | Important | Medium | Low

EPOCH

Sample values: 0 | 1

VERSION

Sample values: 78.6.1 | 4.10.16

RELEASE

Sample values: 9.56.amzn1 | 1.amzn2

ARCH

Sample values: i686 | x86_64

REPOSITORY

Sample values: Core | Updates

ADVISORY_ID

Sample values: ALAS-2018-1058 | ALAS2-2021-1594

CVE_ID

Sample values: CVE-2018-3615 | CVE-2020-1472

BUGZILLA_ID

Sample values: 1463241

Destructuring the Response

data DescribeAvailablePatchesResponse

See: newDescribeAvailablePatchesResponse smart constructor.

Constructors

DescribeAvailablePatchesResponse'

Fields

nextToken :: Maybe Text

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

patches :: Maybe [Patch]

An array of patches. Each entry in the array is a patch structure.

httpStatus :: Int

The response's http status code.

newDescribeAvailablePatchesResponse

Arguments

:: Int
$sel:httpStatus:DescribeAvailablePatchesResponse'

-> DescribeAvailablePatchesResponse

Create a value of DescribeAvailablePatchesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeAvailablePatches', describeAvailablePatchesResponse_nextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

$sel:patches:DescribeAvailablePatchesResponse', describeAvailablePatchesResponse_patches - An array of patches. Each entry in the array is a patch structure.

$sel:httpStatus:DescribeAvailablePatchesResponse', describeAvailablePatchesResponse_httpStatus - The response's http status code.

Response Lenses

describeAvailablePatchesResponse_nextToken :: Lens' DescribeAvailablePatchesResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

describeAvailablePatchesResponse_patches :: Lens' DescribeAvailablePatchesResponse (Maybe [Patch])

An array of patches. Each entry in the array is a patch structure.

describeAvailablePatchesResponse_httpStatus :: Lens' DescribeAvailablePatchesResponse Int

The response's http status code.