Network.AWS.SSM.DescribeDocumentPermission

Description

Describes the permissions for a Amazon Web Services Systems Manager document (SSM document). If you created the document, you are the owner. If a document is shared, it can either be shared privately (by specifying a user's Amazon Web Services account ID) or publicly (All).

Creating a Request

data DescribeDocumentPermission

See: newDescribeDocumentPermission smart constructor.

Constructors

DescribeDocumentPermission'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

name :: Text

The name of the document for which you are the owner.

permissionType :: DocumentPermissionType

The permission type for the document. The permission type can be Share.

newDescribeDocumentPermission

Arguments

:: Text
$sel:name:DescribeDocumentPermission'

-> DocumentPermissionType
$sel:permissionType:DescribeDocumentPermission'

-> DescribeDocumentPermission

Create a value of DescribeDocumentPermission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeDocumentPermission', describeDocumentPermission_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:maxResults:DescribeDocumentPermission', describeDocumentPermission_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

$sel:name:DescribeDocumentPermission', describeDocumentPermission_name - The name of the document for which you are the owner.

$sel:permissionType:DescribeDocumentPermission', describeDocumentPermission_permissionType - The permission type for the document. The permission type can be Share.

Request Lenses

describeDocumentPermission_nextToken :: Lens' DescribeDocumentPermission (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeDocumentPermission_maxResults :: Lens' DescribeDocumentPermission (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

describeDocumentPermission_name :: Lens' DescribeDocumentPermission Text

The name of the document for which you are the owner.

describeDocumentPermission_permissionType :: Lens' DescribeDocumentPermission DocumentPermissionType

The permission type for the document. The permission type can be Share.

Destructuring the Response

data DescribeDocumentPermissionResponse

See: newDescribeDocumentPermissionResponse smart constructor.

Constructors

DescribeDocumentPermissionResponse'

Fields

accountIds :: Maybe [Text]

The account IDs that have permission to use this document. The ID can be either an Amazon Web Services account or All.

accountSharingInfoList :: Maybe [AccountSharingInfo]

A list of Amazon Web Services accounts where the current document is shared and the version shared with each account.

nextToken :: Maybe Text

The token for the next set of items to return. Use this token to get the next set of results.

httpStatus :: Int

The response's http status code.

newDescribeDocumentPermissionResponse

Arguments

:: Int
$sel:httpStatus:DescribeDocumentPermissionResponse'

-> DescribeDocumentPermissionResponse

Create a value of DescribeDocumentPermissionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountIds:DescribeDocumentPermissionResponse', describeDocumentPermissionResponse_accountIds - The account IDs that have permission to use this document. The ID can be either an Amazon Web Services account or All.

$sel:accountSharingInfoList:DescribeDocumentPermissionResponse', describeDocumentPermissionResponse_accountSharingInfoList - A list of Amazon Web Services accounts where the current document is shared and the version shared with each account.

$sel:nextToken:DescribeDocumentPermission', describeDocumentPermissionResponse_nextToken - The token for the next set of items to return. Use this token to get the next set of results.

$sel:httpStatus:DescribeDocumentPermissionResponse', describeDocumentPermissionResponse_httpStatus - The response's http status code.

Response Lenses

describeDocumentPermissionResponse_accountIds :: Lens' DescribeDocumentPermissionResponse (Maybe [Text])

The account IDs that have permission to use this document. The ID can be either an Amazon Web Services account or All.

describeDocumentPermissionResponse_accountSharingInfoList :: Lens' DescribeDocumentPermissionResponse (Maybe [AccountSharingInfo])

A list of Amazon Web Services accounts where the current document is shared and the version shared with each account.

describeDocumentPermissionResponse_nextToken :: Lens' DescribeDocumentPermissionResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

describeDocumentPermissionResponse_httpStatus :: Lens' DescribeDocumentPermissionResponse Int

The response's http status code.