Network.AWS.SSM.DescribeInstanceInformation

Description

Describes one or more of your instances, including information about the operating system platform, the version of SSM Agent installed on the instance, instance status, and so on.

If you specify one or more instance IDs, it returns information for those instances. If you don't specify instance IDs, it returns information for all your instances. If you specify an instance ID that isn't valid or an instance that you don't own, you receive an error.

The IamRole field for this API operation is the Identity and Access Management (IAM) role assigned to on-premises instances. This call doesn't return the IAM role for EC2 instances.

This operation returns paginated results.

Creating a Request

data DescribeInstanceInformation

See: newDescribeInstanceInformation smart constructor.

Constructors

DescribeInstanceInformation'

Fields

instanceInformationFilterList :: Maybe [InstanceInformationFilter]

This is a legacy method. We recommend that you don't use this method. Instead, use the Filters data type. Filters enables you to return instance information by filtering based on tags applied to managed instances.

Attempting to use InstanceInformationFilterList and Filters leads to an exception error.

filters :: Maybe [InstanceInformationStringFilter]

One or more filters. Use a filter to return a more specific list of instances. You can filter based on tags applied to EC2 instances. Use this Filters data type instead of InstanceInformationFilterList, which is deprecated.

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

newDescribeInstanceInformation :: DescribeInstanceInformation

Create a value of DescribeInstanceInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceInformationFilterList:DescribeInstanceInformation', describeInstanceInformation_instanceInformationFilterList - This is a legacy method. We recommend that you don't use this method. Instead, use the Filters data type. Filters enables you to return instance information by filtering based on tags applied to managed instances.

Attempting to use InstanceInformationFilterList and Filters leads to an exception error.

$sel:filters:DescribeInstanceInformation', describeInstanceInformation_filters - One or more filters. Use a filter to return a more specific list of instances. You can filter based on tags applied to EC2 instances. Use this Filters data type instead of InstanceInformationFilterList, which is deprecated.

$sel:nextToken:DescribeInstanceInformation', describeInstanceInformation_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:maxResults:DescribeInstanceInformation', describeInstanceInformation_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

Request Lenses

describeInstanceInformation_instanceInformationFilterList :: Lens' DescribeInstanceInformation (Maybe [InstanceInformationFilter])

This is a legacy method. We recommend that you don't use this method. Instead, use the Filters data type. Filters enables you to return instance information by filtering based on tags applied to managed instances.

Attempting to use InstanceInformationFilterList and Filters leads to an exception error.

describeInstanceInformation_filters :: Lens' DescribeInstanceInformation (Maybe [InstanceInformationStringFilter])

One or more filters. Use a filter to return a more specific list of instances. You can filter based on tags applied to EC2 instances. Use this Filters data type instead of InstanceInformationFilterList, which is deprecated.

describeInstanceInformation_nextToken :: Lens' DescribeInstanceInformation (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeInstanceInformation_maxResults :: Lens' DescribeInstanceInformation (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

Destructuring the Response

data DescribeInstanceInformationResponse

See: newDescribeInstanceInformationResponse smart constructor.

Constructors

DescribeInstanceInformationResponse'

Fields

nextToken :: Maybe Text

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

instanceInformationList :: Maybe [InstanceInformation]

The instance information list.

httpStatus :: Int

The response's http status code.

newDescribeInstanceInformationResponse

Arguments

:: Int
$sel:httpStatus:DescribeInstanceInformationResponse'

-> DescribeInstanceInformationResponse

Create a value of DescribeInstanceInformationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeInstanceInformation', describeInstanceInformationResponse_nextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

$sel:instanceInformationList:DescribeInstanceInformationResponse', describeInstanceInformationResponse_instanceInformationList - The instance information list.

$sel:httpStatus:DescribeInstanceInformationResponse', describeInstanceInformationResponse_httpStatus - The response's http status code.

Response Lenses

describeInstanceInformationResponse_nextToken :: Lens' DescribeInstanceInformationResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

describeInstanceInformationResponse_instanceInformationList :: Lens' DescribeInstanceInformationResponse (Maybe [InstanceInformation])

The instance information list.

describeInstanceInformationResponse_httpStatus :: Lens' DescribeInstanceInformationResponse Int

The response's http status code.