Network.AWS.SSM.DescribeInstancePatches

Description

Retrieves information about the patches on the specified instance and their state relative to the patch baseline being used for the instance.

This operation returns paginated results.

Creating a Request

data DescribeInstancePatches

See: newDescribeInstancePatches smart constructor.

Constructors

DescribeInstancePatches'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

maxResults :: Maybe Natural

The maximum number of patches to return (per page).

filters :: Maybe [PatchOrchestratorFilter]

Each element in the array is a structure containing a key-value pair.

Supported keys for DescribeInstancePatchesinclude the following:

Classification

Sample values: Security | SecurityUpdates

KBId

Sample values: KB4480056 | java-1.7.0-openjdk.x86_64

Severity

Sample values: Important | Medium | Low

State

Sample values: Installed | InstalledOther | InstalledPendingReboot

instanceId :: Text

The ID of the instance whose patch state information should be retrieved.

newDescribeInstancePatches

Arguments

:: Text
$sel:instanceId:DescribeInstancePatches'

-> DescribeInstancePatches

Create a value of DescribeInstancePatches with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeInstancePatches', describeInstancePatches_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:maxResults:DescribeInstancePatches', describeInstancePatches_maxResults - The maximum number of patches to return (per page).

$sel:filters:DescribeInstancePatches', describeInstancePatches_filters - Each element in the array is a structure containing a key-value pair.

Supported keys for DescribeInstancePatchesinclude the following:

Classification

Sample values: Security | SecurityUpdates

KBId

Sample values: KB4480056 | java-1.7.0-openjdk.x86_64

Severity

Sample values: Important | Medium | Low

State

Sample values: Installed | InstalledOther | InstalledPendingReboot

$sel:instanceId:DescribeInstancePatches', describeInstancePatches_instanceId - The ID of the instance whose patch state information should be retrieved.

Request Lenses

describeInstancePatches_nextToken :: Lens' DescribeInstancePatches (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeInstancePatches_maxResults :: Lens' DescribeInstancePatches (Maybe Natural)

The maximum number of patches to return (per page).

describeInstancePatches_filters :: Lens' DescribeInstancePatches (Maybe [PatchOrchestratorFilter])

Each element in the array is a structure containing a key-value pair.

Supported keys for DescribeInstancePatchesinclude the following:

Classification

Sample values: Security | SecurityUpdates

KBId

Sample values: KB4480056 | java-1.7.0-openjdk.x86_64

Severity

Sample values: Important | Medium | Low

State

Sample values: Installed | InstalledOther | InstalledPendingReboot

describeInstancePatches_instanceId :: Lens' DescribeInstancePatches Text

The ID of the instance whose patch state information should be retrieved.

Destructuring the Response

data DescribeInstancePatchesResponse

See: newDescribeInstancePatchesResponse smart constructor.

Constructors

DescribeInstancePatchesResponse'

Fields

nextToken :: Maybe Text

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

patches :: Maybe [PatchComplianceData]

Each entry in the array is a structure containing:

Title (string)

KBId (string)

Classification (string)

Severity (string)

State (string, such as "INSTALLED" or "FAILED")

InstalledTime (DateTime)

InstalledBy (string)

httpStatus :: Int

The response's http status code.

newDescribeInstancePatchesResponse

Arguments

:: Int
$sel:httpStatus:DescribeInstancePatchesResponse'

-> DescribeInstancePatchesResponse

Create a value of DescribeInstancePatchesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeInstancePatches', describeInstancePatchesResponse_nextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

$sel:patches:DescribeInstancePatchesResponse', describeInstancePatchesResponse_patches - Each entry in the array is a structure containing:

Title (string)

KBId (string)

Classification (string)

Severity (string)

State (string, such as "INSTALLED" or "FAILED")

InstalledTime (DateTime)

InstalledBy (string)

$sel:httpStatus:DescribeInstancePatchesResponse', describeInstancePatchesResponse_httpStatus - The response's http status code.

Response Lenses

describeInstancePatchesResponse_nextToken :: Lens' DescribeInstancePatchesResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

describeInstancePatchesResponse_patches :: Lens' DescribeInstancePatchesResponse (Maybe [PatchComplianceData])

Each entry in the array is a structure containing:

Title (string)

KBId (string)

Classification (string)

Severity (string)

State (string, such as "INSTALLED" or "FAILED")

InstalledTime (DateTime)

InstalledBy (string)

describeInstancePatchesResponse_httpStatus :: Lens' DescribeInstancePatchesResponse Int

The response's http status code.