Network.AWS.SSM.DescribeMaintenanceWindowSchedule

Description

Retrieves information about upcoming executions of a maintenance window.

This operation returns paginated results.

Creating a Request

data DescribeMaintenanceWindowSchedule

See: newDescribeMaintenanceWindowSchedule smart constructor.

Constructors

DescribeMaintenanceWindowSchedule'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

resourceType :: Maybe MaintenanceWindowResourceType

The type of resource you want to retrieve information about. For example, INSTANCE.

targets :: Maybe [Target]

The instance ID or key-value pair to retrieve information about.

windowId :: Maybe Text

The ID of the maintenance window to retrieve information about.

filters :: Maybe [PatchOrchestratorFilter]

Filters used to limit the range of results. For example, you can limit maintenance window executions to only those scheduled before or after a certain date and time.

newDescribeMaintenanceWindowSchedule :: DescribeMaintenanceWindowSchedule

Create a value of DescribeMaintenanceWindowSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeMaintenanceWindowSchedule', describeMaintenanceWindowSchedule_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:maxResults:DescribeMaintenanceWindowSchedule', describeMaintenanceWindowSchedule_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

$sel:resourceType:DescribeMaintenanceWindowSchedule', describeMaintenanceWindowSchedule_resourceType - The type of resource you want to retrieve information about. For example, INSTANCE.

$sel:targets:DescribeMaintenanceWindowSchedule', describeMaintenanceWindowSchedule_targets - The instance ID or key-value pair to retrieve information about.

$sel:windowId:DescribeMaintenanceWindowSchedule', describeMaintenanceWindowSchedule_windowId - The ID of the maintenance window to retrieve information about.

$sel:filters:DescribeMaintenanceWindowSchedule', describeMaintenanceWindowSchedule_filters - Filters used to limit the range of results. For example, you can limit maintenance window executions to only those scheduled before or after a certain date and time.

Request Lenses

describeMaintenanceWindowSchedule_nextToken :: Lens' DescribeMaintenanceWindowSchedule (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeMaintenanceWindowSchedule_maxResults :: Lens' DescribeMaintenanceWindowSchedule (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

describeMaintenanceWindowSchedule_resourceType :: Lens' DescribeMaintenanceWindowSchedule (Maybe MaintenanceWindowResourceType)

The type of resource you want to retrieve information about. For example, INSTANCE.

describeMaintenanceWindowSchedule_targets :: Lens' DescribeMaintenanceWindowSchedule (Maybe [Target])

The instance ID or key-value pair to retrieve information about.

describeMaintenanceWindowSchedule_windowId :: Lens' DescribeMaintenanceWindowSchedule (Maybe Text)

The ID of the maintenance window to retrieve information about.

describeMaintenanceWindowSchedule_filters :: Lens' DescribeMaintenanceWindowSchedule (Maybe [PatchOrchestratorFilter])

Filters used to limit the range of results. For example, you can limit maintenance window executions to only those scheduled before or after a certain date and time.

Destructuring the Response

data DescribeMaintenanceWindowScheduleResponse

See: newDescribeMaintenanceWindowScheduleResponse smart constructor.

Constructors

DescribeMaintenanceWindowScheduleResponse'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You use this token in the next call.)

scheduledWindowExecutions :: Maybe [ScheduledWindowExecution]

Information about maintenance window executions scheduled for the specified time range.

httpStatus :: Int

The response's http status code.

newDescribeMaintenanceWindowScheduleResponse

Arguments

:: Int
$sel:httpStatus:DescribeMaintenanceWindowScheduleResponse'

-> DescribeMaintenanceWindowScheduleResponse

Create a value of DescribeMaintenanceWindowScheduleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeMaintenanceWindowSchedule', describeMaintenanceWindowScheduleResponse_nextToken - The token for the next set of items to return. (You use this token in the next call.)

$sel:scheduledWindowExecutions:DescribeMaintenanceWindowScheduleResponse', describeMaintenanceWindowScheduleResponse_scheduledWindowExecutions - Information about maintenance window executions scheduled for the specified time range.

$sel:httpStatus:DescribeMaintenanceWindowScheduleResponse', describeMaintenanceWindowScheduleResponse_httpStatus - The response's http status code.

Response Lenses

describeMaintenanceWindowScheduleResponse_nextToken :: Lens' DescribeMaintenanceWindowScheduleResponse (Maybe Text)

The token for the next set of items to return. (You use this token in the next call.)

describeMaintenanceWindowScheduleResponse_scheduledWindowExecutions :: Lens' DescribeMaintenanceWindowScheduleResponse (Maybe [ScheduledWindowExecution])

Information about maintenance window executions scheduled for the specified time range.

describeMaintenanceWindowScheduleResponse_httpStatus :: Lens' DescribeMaintenanceWindowScheduleResponse Int

The response's http status code.