Network.AWS.SSM.DescribeMaintenanceWindowsForTarget

Description

Retrieves information about the maintenance window targets or tasks that an instance is associated with.

This operation returns paginated results.

Creating a Request

data DescribeMaintenanceWindowsForTarget

See: newDescribeMaintenanceWindowsForTarget smart constructor.

Constructors

DescribeMaintenanceWindowsForTarget'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

targets :: [Target]

The instance ID or key-value pair to retrieve information about.

resourceType :: MaintenanceWindowResourceType

The type of resource you want to retrieve information about. For example, INSTANCE.

newDescribeMaintenanceWindowsForTarget

Arguments

:: MaintenanceWindowResourceType
$sel:resourceType:DescribeMaintenanceWindowsForTarget'

-> DescribeMaintenanceWindowsForTarget

Create a value of DescribeMaintenanceWindowsForTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeMaintenanceWindowsForTarget', describeMaintenanceWindowsForTarget_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:maxResults:DescribeMaintenanceWindowsForTarget', describeMaintenanceWindowsForTarget_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

$sel:targets:DescribeMaintenanceWindowsForTarget', describeMaintenanceWindowsForTarget_targets - The instance ID or key-value pair to retrieve information about.

$sel:resourceType:DescribeMaintenanceWindowsForTarget', describeMaintenanceWindowsForTarget_resourceType - The type of resource you want to retrieve information about. For example, INSTANCE.

Request Lenses

describeMaintenanceWindowsForTarget_nextToken :: Lens' DescribeMaintenanceWindowsForTarget (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeMaintenanceWindowsForTarget_maxResults :: Lens' DescribeMaintenanceWindowsForTarget (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

describeMaintenanceWindowsForTarget_targets :: Lens' DescribeMaintenanceWindowsForTarget [Target]

The instance ID or key-value pair to retrieve information about.

describeMaintenanceWindowsForTarget_resourceType :: Lens' DescribeMaintenanceWindowsForTarget MaintenanceWindowResourceType

The type of resource you want to retrieve information about. For example, INSTANCE.

Destructuring the Response

data DescribeMaintenanceWindowsForTargetResponse

See: newDescribeMaintenanceWindowsForTargetResponse smart constructor.

Constructors

DescribeMaintenanceWindowsForTargetResponse'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You use this token in the next call.)

windowIdentities :: Maybe [MaintenanceWindowIdentityForTarget]

Information about the maintenance window targets and tasks an instance is associated with.

httpStatus :: Int

The response's http status code.

newDescribeMaintenanceWindowsForTargetResponse

Arguments

:: Int
$sel:httpStatus:DescribeMaintenanceWindowsForTargetResponse'

-> DescribeMaintenanceWindowsForTargetResponse

Create a value of DescribeMaintenanceWindowsForTargetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeMaintenanceWindowsForTarget', describeMaintenanceWindowsForTargetResponse_nextToken - The token for the next set of items to return. (You use this token in the next call.)

$sel:windowIdentities:DescribeMaintenanceWindowsForTargetResponse', describeMaintenanceWindowsForTargetResponse_windowIdentities - Information about the maintenance window targets and tasks an instance is associated with.

$sel:httpStatus:DescribeMaintenanceWindowsForTargetResponse', describeMaintenanceWindowsForTargetResponse_httpStatus - The response's http status code.

Response Lenses

describeMaintenanceWindowsForTargetResponse_nextToken :: Lens' DescribeMaintenanceWindowsForTargetResponse (Maybe Text)

The token for the next set of items to return. (You use this token in the next call.)

describeMaintenanceWindowsForTargetResponse_windowIdentities :: Lens' DescribeMaintenanceWindowsForTargetResponse (Maybe [MaintenanceWindowIdentityForTarget])

Information about the maintenance window targets and tasks an instance is associated with.

describeMaintenanceWindowsForTargetResponse_httpStatus :: Lens' DescribeMaintenanceWindowsForTargetResponse Int

The response's http status code.