Network.AWS.SSM.DescribeOpsItems

Description

Query a set of OpsItems. You must have permission in Identity and Access Management (IAM) to query a list of OpsItems. For more information, see Getting started with OpsCenter in the Amazon Web Services Systems Manager User Guide.

Operations engineers and IT professionals use Amazon Web Services Systems Manager OpsCenter to view, investigate, and remediate operational issues impacting the performance and health of their Amazon Web Services resources. For more information, see OpsCenter in the Amazon Web Services Systems Manager User Guide.

This operation returns paginated results.

Creating a Request

data DescribeOpsItems

See: newDescribeOpsItems smart constructor.

Constructors

DescribeOpsItems'

Fields

nextToken :: Maybe Text

A token to start the list. Use this token to get the next set of results.

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

opsItemFilters :: Maybe [OpsItemFilter]

One or more filters to limit the response.

Key: CreatedTime

Operations: GreaterThan, LessThan

Key: LastModifiedBy

Operations: Contains, Equals

Key: LastModifiedTime

Operations: GreaterThan, LessThan

Key: Priority

Operations: Equals

Key: Source

Operations: Contains, Equals

Key: Status

Operations: Equals

Key: Title*

Operations: Equals,Contains

Key: OperationalData**

Operations: Equals

Key: OperationalDataKey

Operations: Equals

Key: OperationalDataValue

Operations: Equals, Contains

Key: OpsItemId

Operations: Equals

Key: ResourceId

Operations: Contains

Key: AutomationId

Operations: Equals

The Equals operator for Title matches the first 100 characters. If you specify more than 100 characters, they system returns an error that the filter value exceeds the length limit.

*If you filter the response by using the OperationalData operator, specify a key-value pair by using the following JSON format: {"key":"key_name","value":"a_value"}

newDescribeOpsItems :: DescribeOpsItems

Create a value of DescribeOpsItems with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeOpsItems', describeOpsItems_nextToken - A token to start the list. Use this token to get the next set of results.

$sel:maxResults:DescribeOpsItems', describeOpsItems_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

$sel:opsItemFilters:DescribeOpsItems', describeOpsItems_opsItemFilters - One or more filters to limit the response.

Key: CreatedTime

Operations: GreaterThan, LessThan

Key: LastModifiedBy

Operations: Contains, Equals

Key: LastModifiedTime

Operations: GreaterThan, LessThan

Key: Priority

Operations: Equals

Key: Source

Operations: Contains, Equals

Key: Status

Operations: Equals

Key: Title*

Operations: Equals,Contains

Key: OperationalData**

Operations: Equals

Key: OperationalDataKey

Operations: Equals

Key: OperationalDataValue

Operations: Equals, Contains

Key: OpsItemId

Operations: Equals

Key: ResourceId

Operations: Contains

Key: AutomationId

Operations: Equals

The Equals operator for Title matches the first 100 characters. If you specify more than 100 characters, they system returns an error that the filter value exceeds the length limit.

*If you filter the response by using the OperationalData operator, specify a key-value pair by using the following JSON format: {"key":"key_name","value":"a_value"}

Request Lenses

describeOpsItems_nextToken :: Lens' DescribeOpsItems (Maybe Text)

A token to start the list. Use this token to get the next set of results.

describeOpsItems_maxResults :: Lens' DescribeOpsItems (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

describeOpsItems_opsItemFilters :: Lens' DescribeOpsItems (Maybe [OpsItemFilter])

One or more filters to limit the response.

Key: CreatedTime

Operations: GreaterThan, LessThan

Key: LastModifiedBy

Operations: Contains, Equals

Key: LastModifiedTime

Operations: GreaterThan, LessThan

Key: Priority

Operations: Equals

Key: Source

Operations: Contains, Equals

Key: Status

Operations: Equals

Key: Title*

Operations: Equals,Contains

Key: OperationalData**

Operations: Equals

Key: OperationalDataKey

Operations: Equals

Key: OperationalDataValue

Operations: Equals, Contains

Key: OpsItemId

Operations: Equals

Key: ResourceId

Operations: Contains

Key: AutomationId

Operations: Equals

The Equals operator for Title matches the first 100 characters. If you specify more than 100 characters, they system returns an error that the filter value exceeds the length limit.

*If you filter the response by using the OperationalData operator, specify a key-value pair by using the following JSON format: {"key":"key_name","value":"a_value"}

Destructuring the Response

data DescribeOpsItemsResponse

See: newDescribeOpsItemsResponse smart constructor.

Constructors

DescribeOpsItemsResponse'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. Use this token to get the next set of results.

opsItemSummaries :: Maybe [OpsItemSummary]

A list of OpsItems.

httpStatus :: Int

The response's http status code.

newDescribeOpsItemsResponse

Arguments

:: Int
$sel:httpStatus:DescribeOpsItemsResponse'

-> DescribeOpsItemsResponse

Create a value of DescribeOpsItemsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeOpsItems', describeOpsItemsResponse_nextToken - The token for the next set of items to return. Use this token to get the next set of results.

$sel:opsItemSummaries:DescribeOpsItemsResponse', describeOpsItemsResponse_opsItemSummaries - A list of OpsItems.

$sel:httpStatus:DescribeOpsItemsResponse', describeOpsItemsResponse_httpStatus - The response's http status code.

Response Lenses

describeOpsItemsResponse_nextToken :: Lens' DescribeOpsItemsResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

describeOpsItemsResponse_opsItemSummaries :: Lens' DescribeOpsItemsResponse (Maybe [OpsItemSummary])

A list of OpsItems.

describeOpsItemsResponse_httpStatus :: Lens' DescribeOpsItemsResponse Int

The response's http status code.