Network.AWS.SSM.DescribeParameters

Description

Get information about a parameter.

Request results are returned on a best-effort basis. If you specify MaxResults in the request, the response includes information up to the limit specified. The number of items returned, however, can be between zero and the value of MaxResults. If the service reaches an internal limit while processing the results, it stops the operation and returns the matching values up to that point and a NextToken. You can specify the NextToken in a subsequent call to get the next set of results.

If you change the KMS key alias for the KMS key used to encrypt a parameter, then you must also update the key alias the parameter uses to reference KMS. Otherwise, DescribeParameters retrieves whatever the original key alias was referencing.

This operation returns paginated results.

Creating a Request

data DescribeParameters

See: newDescribeParameters smart constructor.

Constructors

DescribeParameters'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

parameterFilters :: Maybe [ParameterStringFilter]

Filters to limit the request results.

filters :: Maybe [ParametersFilter]

This data type is deprecated. Instead, use ParameterFilters.

newDescribeParameters :: DescribeParameters

Create a value of DescribeParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeParameters', describeParameters_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:maxResults:DescribeParameters', describeParameters_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

$sel:parameterFilters:DescribeParameters', describeParameters_parameterFilters - Filters to limit the request results.

$sel:filters:DescribeParameters', describeParameters_filters - This data type is deprecated. Instead, use ParameterFilters.

Request Lenses

describeParameters_nextToken :: Lens' DescribeParameters (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeParameters_maxResults :: Lens' DescribeParameters (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

describeParameters_parameterFilters :: Lens' DescribeParameters (Maybe [ParameterStringFilter])

Filters to limit the request results.

describeParameters_filters :: Lens' DescribeParameters (Maybe [ParametersFilter])

This data type is deprecated. Instead, use ParameterFilters.

Destructuring the Response

data DescribeParametersResponse

See: newDescribeParametersResponse smart constructor.

Constructors

DescribeParametersResponse'

Fields

nextToken :: Maybe Text

The token to use when requesting the next set of items.

parameters :: Maybe [ParameterMetadata]

Parameters returned by the request.

httpStatus :: Int

The response's http status code.

newDescribeParametersResponse

Arguments

:: Int
$sel:httpStatus:DescribeParametersResponse'

-> DescribeParametersResponse

Create a value of DescribeParametersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeParameters', describeParametersResponse_nextToken - The token to use when requesting the next set of items.

$sel:parameters:DescribeParametersResponse', describeParametersResponse_parameters - Parameters returned by the request.

$sel:httpStatus:DescribeParametersResponse', describeParametersResponse_httpStatus - The response's http status code.

Response Lenses

describeParametersResponse_nextToken :: Lens' DescribeParametersResponse (Maybe Text)

The token to use when requesting the next set of items.

describeParametersResponse_parameters :: Lens' DescribeParametersResponse (Maybe [ParameterMetadata])

Parameters returned by the request.

describeParametersResponse_httpStatus :: Lens' DescribeParametersResponse Int

The response's http status code.