Network.AWS.SSM.DescribePatchBaselines

Description

Lists the patch baselines in your Amazon Web Services account.

This operation returns paginated results.

Creating a Request

data DescribePatchBaselines

See: newDescribePatchBaselines smart constructor.

Constructors

DescribePatchBaselines'

Fields

filters :: Maybe [PatchOrchestratorFilter]

Each element in the array is a structure containing a key-value pair.

Supported keys for DescribePatchBaselines include the following:

NAME_PREFIX

Sample values: AWS- | My-

OWNER

Sample values: AWS | Self

OPERATING_SYSTEM

Sample values: AMAZON_LINUX | SUSE | WINDOWS

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

maxResults :: Maybe Natural

The maximum number of patch baselines to return (per page).

newDescribePatchBaselines :: DescribePatchBaselines

Create a value of DescribePatchBaselines with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribePatchBaselines', describePatchBaselines_filters - Each element in the array is a structure containing a key-value pair.

Supported keys for DescribePatchBaselines include the following:

NAME_PREFIX

Sample values: AWS- | My-

OWNER

Sample values: AWS | Self

OPERATING_SYSTEM

Sample values: AMAZON_LINUX | SUSE | WINDOWS

$sel:nextToken:DescribePatchBaselines', describePatchBaselines_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:maxResults:DescribePatchBaselines', describePatchBaselines_maxResults - The maximum number of patch baselines to return (per page).

Request Lenses

describePatchBaselines_filters :: Lens' DescribePatchBaselines (Maybe [PatchOrchestratorFilter])

Each element in the array is a structure containing a key-value pair.

Supported keys for DescribePatchBaselines include the following:

NAME_PREFIX

Sample values: AWS- | My-

OWNER

Sample values: AWS | Self

OPERATING_SYSTEM

Sample values: AMAZON_LINUX | SUSE | WINDOWS

describePatchBaselines_nextToken :: Lens' DescribePatchBaselines (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describePatchBaselines_maxResults :: Lens' DescribePatchBaselines (Maybe Natural)

The maximum number of patch baselines to return (per page).

Destructuring the Response

data DescribePatchBaselinesResponse

See: newDescribePatchBaselinesResponse smart constructor.

Constructors

DescribePatchBaselinesResponse'

Fields

baselineIdentities :: Maybe [PatchBaselineIdentity]

An array of PatchBaselineIdentity elements.

nextToken :: Maybe Text

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

httpStatus :: Int

The response's http status code.

newDescribePatchBaselinesResponse

Arguments

:: Int
$sel:httpStatus:DescribePatchBaselinesResponse'

-> DescribePatchBaselinesResponse

Create a value of DescribePatchBaselinesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baselineIdentities:DescribePatchBaselinesResponse', describePatchBaselinesResponse_baselineIdentities - An array of PatchBaselineIdentity elements.

$sel:nextToken:DescribePatchBaselines', describePatchBaselinesResponse_nextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

$sel:httpStatus:DescribePatchBaselinesResponse', describePatchBaselinesResponse_httpStatus - The response's http status code.

Response Lenses

describePatchBaselinesResponse_baselineIdentities :: Lens' DescribePatchBaselinesResponse (Maybe [PatchBaselineIdentity])

An array of PatchBaselineIdentity elements.

describePatchBaselinesResponse_nextToken :: Lens' DescribePatchBaselinesResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

describePatchBaselinesResponse_httpStatus :: Lens' DescribePatchBaselinesResponse Int

The response's http status code.