Network.AWS.SSM.DescribePatchGroupState

Description

Returns high-level aggregated patch compliance state information for a patch group.

Creating a Request

data DescribePatchGroupState

See: newDescribePatchGroupState smart constructor.

Constructors

DescribePatchGroupState'

Fields

patchGroup :: Text

The name of the patch group whose patch snapshot should be retrieved.

newDescribePatchGroupState

Arguments

:: Text
$sel:patchGroup:DescribePatchGroupState'

-> DescribePatchGroupState

Create a value of DescribePatchGroupState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:patchGroup:DescribePatchGroupState', describePatchGroupState_patchGroup - The name of the patch group whose patch snapshot should be retrieved.

Request Lenses

describePatchGroupState_patchGroup :: Lens' DescribePatchGroupState Text

The name of the patch group whose patch snapshot should be retrieved.

Destructuring the Response

data DescribePatchGroupStateResponse

See: newDescribePatchGroupStateResponse smart constructor.

Constructors

DescribePatchGroupStateResponse'

Fields

instancesWithMissingPatches :: Maybe Int

The number of instances with missing patches from the patch baseline.

instancesWithInstalledOtherPatches :: Maybe Int

The number of instances with patches installed that aren't defined in the patch baseline.

instancesWithNotApplicablePatches :: Maybe Int

The number of instances with patches that aren't applicable.

instancesWithInstalledPatches :: Maybe Int

The number of instances with installed patches.

instancesWithCriticalNonCompliantPatches :: Maybe Int

The number of instances where patches that are specified as Critical for compliance reporting in the patch baseline aren't installed. These patches might be missing, have failed installation, were rejected, or were installed but awaiting a required instance reboot. The status of these instances is NON_COMPLIANT.

instancesWithSecurityNonCompliantPatches :: Maybe Int

The number of instances where patches that are specified as Security in a patch advisory aren't installed. These patches might be missing, have failed installation, were rejected, or were installed but awaiting a required instance reboot. The status of these instances is NON_COMPLIANT.

instancesWithInstalledRejectedPatches :: Maybe Int

The number of instances with patches installed that are specified in a RejectedPatches list. Patches with a status of INSTALLED_REJECTED were typically installed before they were added to a RejectedPatches list.

If ALLOW_AS_DEPENDENCY is the specified option for RejectedPatchesAction, the value of InstancesWithInstalledRejectedPatches will always be 0 (zero).

instancesWithInstalledPendingRebootPatches :: Maybe Int

The number of instances with patches installed by Patch Manager that haven't been rebooted after the patch installation. The status of these instances is NON_COMPLIANT.

instancesWithOtherNonCompliantPatches :: Maybe Int

The number of instances with patches installed that are specified as other than Critical or Security but aren't compliant with the patch baseline. The status of these instances is NON_COMPLIANT.

instancesWithUnreportedNotApplicablePatches :: Maybe Int

The number of instances with NotApplicable patches beyond the supported limit, which aren't reported by name to Inventory. Inventory is a capability of Amazon Web Services Systems Manager.

instances :: Maybe Int

The number of instances in the patch group.

instancesWithFailedPatches :: Maybe Int

The number of instances with patches from the patch baseline that failed to install.

httpStatus :: Int

The response's http status code.

newDescribePatchGroupStateResponse

Arguments

:: Int
$sel:httpStatus:DescribePatchGroupStateResponse'

-> DescribePatchGroupStateResponse

Create a value of DescribePatchGroupStateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instancesWithMissingPatches:DescribePatchGroupStateResponse', describePatchGroupStateResponse_instancesWithMissingPatches - The number of instances with missing patches from the patch baseline.

$sel:instancesWithInstalledOtherPatches:DescribePatchGroupStateResponse', describePatchGroupStateResponse_instancesWithInstalledOtherPatches - The number of instances with patches installed that aren't defined in the patch baseline.

$sel:instancesWithNotApplicablePatches:DescribePatchGroupStateResponse', describePatchGroupStateResponse_instancesWithNotApplicablePatches - The number of instances with patches that aren't applicable.

$sel:instancesWithInstalledPatches:DescribePatchGroupStateResponse', describePatchGroupStateResponse_instancesWithInstalledPatches - The number of instances with installed patches.

$sel:instancesWithCriticalNonCompliantPatches:DescribePatchGroupStateResponse', describePatchGroupStateResponse_instancesWithCriticalNonCompliantPatches - The number of instances where patches that are specified as Critical for compliance reporting in the patch baseline aren't installed. These patches might be missing, have failed installation, were rejected, or were installed but awaiting a required instance reboot. The status of these instances is NON_COMPLIANT.

$sel:instancesWithSecurityNonCompliantPatches:DescribePatchGroupStateResponse', describePatchGroupStateResponse_instancesWithSecurityNonCompliantPatches - The number of instances where patches that are specified as Security in a patch advisory aren't installed. These patches might be missing, have failed installation, were rejected, or were installed but awaiting a required instance reboot. The status of these instances is NON_COMPLIANT.

$sel:instancesWithInstalledRejectedPatches:DescribePatchGroupStateResponse', describePatchGroupStateResponse_instancesWithInstalledRejectedPatches - The number of instances with patches installed that are specified in a RejectedPatches list. Patches with a status of INSTALLED_REJECTED were typically installed before they were added to a RejectedPatches list.

If ALLOW_AS_DEPENDENCY is the specified option for RejectedPatchesAction, the value of InstancesWithInstalledRejectedPatches will always be 0 (zero).

$sel:instancesWithInstalledPendingRebootPatches:DescribePatchGroupStateResponse', describePatchGroupStateResponse_instancesWithInstalledPendingRebootPatches - The number of instances with patches installed by Patch Manager that haven't been rebooted after the patch installation. The status of these instances is NON_COMPLIANT.

$sel:instancesWithOtherNonCompliantPatches:DescribePatchGroupStateResponse', describePatchGroupStateResponse_instancesWithOtherNonCompliantPatches - The number of instances with patches installed that are specified as other than Critical or Security but aren't compliant with the patch baseline. The status of these instances is NON_COMPLIANT.

$sel:instancesWithUnreportedNotApplicablePatches:DescribePatchGroupStateResponse', describePatchGroupStateResponse_instancesWithUnreportedNotApplicablePatches - The number of instances with NotApplicable patches beyond the supported limit, which aren't reported by name to Inventory. Inventory is a capability of Amazon Web Services Systems Manager.

$sel:instances:DescribePatchGroupStateResponse', describePatchGroupStateResponse_instances - The number of instances in the patch group.

$sel:instancesWithFailedPatches:DescribePatchGroupStateResponse', describePatchGroupStateResponse_instancesWithFailedPatches - The number of instances with patches from the patch baseline that failed to install.

$sel:httpStatus:DescribePatchGroupStateResponse', describePatchGroupStateResponse_httpStatus - The response's http status code.

Response Lenses

describePatchGroupStateResponse_instancesWithMissingPatches :: Lens' DescribePatchGroupStateResponse (Maybe Int)

The number of instances with missing patches from the patch baseline.

describePatchGroupStateResponse_instancesWithInstalledOtherPatches :: Lens' DescribePatchGroupStateResponse (Maybe Int)

The number of instances with patches installed that aren't defined in the patch baseline.

describePatchGroupStateResponse_instancesWithNotApplicablePatches :: Lens' DescribePatchGroupStateResponse (Maybe Int)

The number of instances with patches that aren't applicable.

describePatchGroupStateResponse_instancesWithInstalledPatches :: Lens' DescribePatchGroupStateResponse (Maybe Int)

The number of instances with installed patches.

describePatchGroupStateResponse_instancesWithCriticalNonCompliantPatches :: Lens' DescribePatchGroupStateResponse (Maybe Int)

The number of instances where patches that are specified as Critical for compliance reporting in the patch baseline aren't installed. These patches might be missing, have failed installation, were rejected, or were installed but awaiting a required instance reboot. The status of these instances is NON_COMPLIANT.

describePatchGroupStateResponse_instancesWithSecurityNonCompliantPatches :: Lens' DescribePatchGroupStateResponse (Maybe Int)

The number of instances where patches that are specified as Security in a patch advisory aren't installed. These patches might be missing, have failed installation, were rejected, or were installed but awaiting a required instance reboot. The status of these instances is NON_COMPLIANT.

describePatchGroupStateResponse_instancesWithInstalledRejectedPatches :: Lens' DescribePatchGroupStateResponse (Maybe Int)

The number of instances with patches installed that are specified in a RejectedPatches list. Patches with a status of INSTALLED_REJECTED were typically installed before they were added to a RejectedPatches list.

If ALLOW_AS_DEPENDENCY is the specified option for RejectedPatchesAction, the value of InstancesWithInstalledRejectedPatches will always be 0 (zero).

describePatchGroupStateResponse_instancesWithInstalledPendingRebootPatches :: Lens' DescribePatchGroupStateResponse (Maybe Int)

The number of instances with patches installed by Patch Manager that haven't been rebooted after the patch installation. The status of these instances is NON_COMPLIANT.

describePatchGroupStateResponse_instancesWithOtherNonCompliantPatches :: Lens' DescribePatchGroupStateResponse (Maybe Int)

The number of instances with patches installed that are specified as other than Critical or Security but aren't compliant with the patch baseline. The status of these instances is NON_COMPLIANT.

describePatchGroupStateResponse_instancesWithUnreportedNotApplicablePatches :: Lens' DescribePatchGroupStateResponse (Maybe Int)

The number of instances with NotApplicable patches beyond the supported limit, which aren't reported by name to Inventory. Inventory is a capability of Amazon Web Services Systems Manager.

describePatchGroupStateResponse_instances :: Lens' DescribePatchGroupStateResponse (Maybe Int)

The number of instances in the patch group.

describePatchGroupStateResponse_instancesWithFailedPatches :: Lens' DescribePatchGroupStateResponse (Maybe Int)

The number of instances with patches from the patch baseline that failed to install.

describePatchGroupStateResponse_httpStatus :: Lens' DescribePatchGroupStateResponse Int

The response's http status code.