Network.AWS.SSM.GetCommandInvocation

Description

Returns detailed information about command execution for an invocation or plugin.

GetCommandInvocation only gives the execution status of a plugin in a document. To get the command execution status on a specific instance, use ListCommandInvocations. To get the command execution status across instances, use ListCommands.

Creating a Request

data GetCommandInvocation

See: newGetCommandInvocation smart constructor.

Constructors

GetCommandInvocation'

Fields

pluginName :: Maybe Text

The name of the plugin for which you want detailed results. If the document contains only one plugin, you can omit the name and details for that plugin. If the document contains more than one plugin, you must specify the name of the plugin for which you want to view details.

Plugin names are also referred to as step names in Systems Manager documents (SSM documents). For example, aws:RunShellScript is a plugin.

To find the PluginName, check the document content and find the name of the plugin. Alternatively, use ListCommandInvocations with the CommandId and Details parameters. The PluginName is the Name attribute of the CommandPlugin object in the CommandPlugins list.

commandId :: Text

(Required) The parent command ID of the invocation plugin.

instanceId :: Text

(Required) The ID of the managed instance targeted by the command. A managed instance can be an Amazon Elastic Compute Cloud (Amazon EC2) instance or an instance in your hybrid environment that is configured for Amazon Web Services Systems Manager.

newGetCommandInvocation

Arguments

:: Text
$sel:commandId:GetCommandInvocation'

-> Text
$sel:instanceId:GetCommandInvocation'

-> GetCommandInvocation

Create a value of GetCommandInvocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pluginName:GetCommandInvocation', getCommandInvocation_pluginName - The name of the plugin for which you want detailed results. If the document contains only one plugin, you can omit the name and details for that plugin. If the document contains more than one plugin, you must specify the name of the plugin for which you want to view details.

Plugin names are also referred to as step names in Systems Manager documents (SSM documents). For example, aws:RunShellScript is a plugin.

To find the PluginName, check the document content and find the name of the plugin. Alternatively, use ListCommandInvocations with the CommandId and Details parameters. The PluginName is the Name attribute of the CommandPlugin object in the CommandPlugins list.

$sel:commandId:GetCommandInvocation', getCommandInvocation_commandId - (Required) The parent command ID of the invocation plugin.

$sel:instanceId:GetCommandInvocation', getCommandInvocation_instanceId - (Required) The ID of the managed instance targeted by the command. A managed instance can be an Amazon Elastic Compute Cloud (Amazon EC2) instance or an instance in your hybrid environment that is configured for Amazon Web Services Systems Manager.

Request Lenses

getCommandInvocation_pluginName :: Lens' GetCommandInvocation (Maybe Text)

The name of the plugin for which you want detailed results. If the document contains only one plugin, you can omit the name and details for that plugin. If the document contains more than one plugin, you must specify the name of the plugin for which you want to view details.

Plugin names are also referred to as step names in Systems Manager documents (SSM documents). For example, aws:RunShellScript is a plugin.

To find the PluginName, check the document content and find the name of the plugin. Alternatively, use ListCommandInvocations with the CommandId and Details parameters. The PluginName is the Name attribute of the CommandPlugin object in the CommandPlugins list.

getCommandInvocation_commandId :: Lens' GetCommandInvocation Text

(Required) The parent command ID of the invocation plugin.

getCommandInvocation_instanceId :: Lens' GetCommandInvocation Text

(Required) The ID of the managed instance targeted by the command. A managed instance can be an Amazon Elastic Compute Cloud (Amazon EC2) instance or an instance in your hybrid environment that is configured for Amazon Web Services Systems Manager.

Destructuring the Response

data GetCommandInvocationResponse

See: newGetCommandInvocationResponse smart constructor.

Constructors

GetCommandInvocationResponse'

Fields

instanceId :: Maybe Text

The ID of the managed instance targeted by the command. A managed instance can be an EC2 instance or an instance in your hybrid environment that is configured for Systems Manager.

status :: Maybe CommandInvocationStatus

The status of this invocation plugin. This status can be different than StatusDetails.

standardErrorContent :: Maybe Text

The first 8,000 characters written by the plugin to stderr. If the command hasn't finished running, then this string is empty.

cloudWatchOutputConfig :: Maybe CloudWatchOutputConfig

Amazon CloudWatch Logs information where Systems Manager sent the command output.

executionElapsedTime :: Maybe Text

Duration since ExecutionStartDateTime.

documentName :: Maybe Text

The name of the document that was run. For example, AWS-RunShellScript.

standardErrorUrl :: Maybe Text

The URL for the complete text written by the plugin to stderr. If the command hasn't finished running, then this string is empty.

executionStartDateTime :: Maybe Text

The date and time the plugin started running. Date and time are written in ISO 8601 format. For example, June 7, 2017 is represented as 2017-06-7. The following sample Amazon Web Services CLI command uses the InvokedBefore filter.

aws ssm list-commands --filters key=InvokedBefore,value=2017-06-07T00:00:00Z

If the plugin hasn't started to run, the string is empty.

responseCode :: Maybe Int

The error level response code for the plugin script. If the response code is -1, then the command hasn't started running on the instance, or it wasn't received by the instance.

statusDetails :: Maybe Text

A detailed status of the command execution for an invocation. StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Understanding command statuses in the Amazon Web Services Systems Manager User Guide. StatusDetails can be one of the following values:

Pending: The command hasn't been sent to the instance.

In Progress: The command has been sent to the instance but hasn't reached a terminal state.

Delayed: The system attempted to send the command to the target, but the target wasn't available. The instance might not be available because of network issues, because the instance was stopped, or for similar reasons. The system will try to send the command again.

Success: The command or plugin ran successfully. This is a terminal state.

Delivery Timed Out: The command wasn't delivered to the instance before the delivery timeout expired. Delivery timeouts don't count against the parent command's MaxErrors limit, but they do contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Execution Timed Out: The command started to run on the instance, but the execution wasn't complete before the timeout expired. Execution timeouts count against the MaxErrors limit of the parent command. This is a terminal state.

Failed: The command wasn't run successfully on the instance. For a plugin, this indicates that the result code wasn't zero. For a command invocation, this indicates that the result code for one or more plugins wasn't zero. Invocation failures count against the MaxErrors limit of the parent command. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Undeliverable: The command can't be delivered to the instance. The instance might not exist or might not be responding. Undeliverable invocations don't count against the parent command's MaxErrors limit and don't contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Terminated: The parent command exceeded its MaxErrors limit and subsequent command invocations were canceled by the system. This is a terminal state.

executionEndDateTime :: Maybe Text

The date and time the plugin finished running. Date and time are written in ISO 8601 format. For example, June 7, 2017 is represented as 2017-06-7. The following sample Amazon Web Services CLI command uses the InvokedAfter filter.

aws ssm list-commands --filters key=InvokedAfter,value=2017-06-07T00:00:00Z

If the plugin hasn't started to run, the string is empty.

standardOutputUrl :: Maybe Text

The URL for the complete text written by the plugin to stdout in Amazon Simple Storage Service (Amazon S3). If an S3 bucket wasn't specified, then this string is empty.

commandId :: Maybe Text

The parent command ID of the invocation plugin.

documentVersion :: Maybe Text

The Systems Manager document (SSM document) version used in the request.

standardOutputContent :: Maybe Text

The first 24,000 characters written by the plugin to stdout. If the command hasn't finished running, if ExecutionStatus is neither Succeeded nor Failed, then this string is empty.

comment :: Maybe Text

The comment text for the command.

pluginName :: Maybe Text

The name of the plugin, or step name, for which details are reported. For example, aws:RunShellScript is a plugin.

httpStatus :: Int

The response's http status code.

newGetCommandInvocationResponse

Arguments

:: Int
$sel:httpStatus:GetCommandInvocationResponse'

-> GetCommandInvocationResponse

Create a value of GetCommandInvocationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:GetCommandInvocation', getCommandInvocationResponse_instanceId - The ID of the managed instance targeted by the command. A managed instance can be an EC2 instance or an instance in your hybrid environment that is configured for Systems Manager.

$sel:status:GetCommandInvocationResponse', getCommandInvocationResponse_status - The status of this invocation plugin. This status can be different than StatusDetails.

$sel:standardErrorContent:GetCommandInvocationResponse', getCommandInvocationResponse_standardErrorContent - The first 8,000 characters written by the plugin to stderr. If the command hasn't finished running, then this string is empty.

$sel:cloudWatchOutputConfig:GetCommandInvocationResponse', getCommandInvocationResponse_cloudWatchOutputConfig - Amazon CloudWatch Logs information where Systems Manager sent the command output.

$sel:executionElapsedTime:GetCommandInvocationResponse', getCommandInvocationResponse_executionElapsedTime - Duration since ExecutionStartDateTime.

$sel:documentName:GetCommandInvocationResponse', getCommandInvocationResponse_documentName - The name of the document that was run. For example, AWS-RunShellScript.

$sel:standardErrorUrl:GetCommandInvocationResponse', getCommandInvocationResponse_standardErrorUrl - The URL for the complete text written by the plugin to stderr. If the command hasn't finished running, then this string is empty.

$sel:executionStartDateTime:GetCommandInvocationResponse', getCommandInvocationResponse_executionStartDateTime - The date and time the plugin started running. Date and time are written in ISO 8601 format. For example, June 7, 2017 is represented as 2017-06-7. The following sample Amazon Web Services CLI command uses the InvokedBefore filter.

aws ssm list-commands --filters key=InvokedBefore,value=2017-06-07T00:00:00Z

If the plugin hasn't started to run, the string is empty.

$sel:responseCode:GetCommandInvocationResponse', getCommandInvocationResponse_responseCode - The error level response code for the plugin script. If the response code is -1, then the command hasn't started running on the instance, or it wasn't received by the instance.

$sel:statusDetails:GetCommandInvocationResponse', getCommandInvocationResponse_statusDetails - A detailed status of the command execution for an invocation. StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Understanding command statuses in the Amazon Web Services Systems Manager User Guide. StatusDetails can be one of the following values:

Pending: The command hasn't been sent to the instance.

In Progress: The command has been sent to the instance but hasn't reached a terminal state.

Delayed: The system attempted to send the command to the target, but the target wasn't available. The instance might not be available because of network issues, because the instance was stopped, or for similar reasons. The system will try to send the command again.

Success: The command or plugin ran successfully. This is a terminal state.

Delivery Timed Out: The command wasn't delivered to the instance before the delivery timeout expired. Delivery timeouts don't count against the parent command's MaxErrors limit, but they do contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Execution Timed Out: The command started to run on the instance, but the execution wasn't complete before the timeout expired. Execution timeouts count against the MaxErrors limit of the parent command. This is a terminal state.

Failed: The command wasn't run successfully on the instance. For a plugin, this indicates that the result code wasn't zero. For a command invocation, this indicates that the result code for one or more plugins wasn't zero. Invocation failures count against the MaxErrors limit of the parent command. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Undeliverable: The command can't be delivered to the instance. The instance might not exist or might not be responding. Undeliverable invocations don't count against the parent command's MaxErrors limit and don't contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Terminated: The parent command exceeded its MaxErrors limit and subsequent command invocations were canceled by the system. This is a terminal state.

$sel:executionEndDateTime:GetCommandInvocationResponse', getCommandInvocationResponse_executionEndDateTime - The date and time the plugin finished running. Date and time are written in ISO 8601 format. For example, June 7, 2017 is represented as 2017-06-7. The following sample Amazon Web Services CLI command uses the InvokedAfter filter.

aws ssm list-commands --filters key=InvokedAfter,value=2017-06-07T00:00:00Z

If the plugin hasn't started to run, the string is empty.

$sel:standardOutputUrl:GetCommandInvocationResponse', getCommandInvocationResponse_standardOutputUrl - The URL for the complete text written by the plugin to stdout in Amazon Simple Storage Service (Amazon S3). If an S3 bucket wasn't specified, then this string is empty.

$sel:commandId:GetCommandInvocation', getCommandInvocationResponse_commandId - The parent command ID of the invocation plugin.

$sel:documentVersion:GetCommandInvocationResponse', getCommandInvocationResponse_documentVersion - The Systems Manager document (SSM document) version used in the request.

$sel:standardOutputContent:GetCommandInvocationResponse', getCommandInvocationResponse_standardOutputContent - The first 24,000 characters written by the plugin to stdout. If the command hasn't finished running, if ExecutionStatus is neither Succeeded nor Failed, then this string is empty.

$sel:comment:GetCommandInvocationResponse', getCommandInvocationResponse_comment - The comment text for the command.

$sel:pluginName:GetCommandInvocation', getCommandInvocationResponse_pluginName - The name of the plugin, or step name, for which details are reported. For example, aws:RunShellScript is a plugin.

$sel:httpStatus:GetCommandInvocationResponse', getCommandInvocationResponse_httpStatus - The response's http status code.

Response Lenses

getCommandInvocationResponse_instanceId :: Lens' GetCommandInvocationResponse (Maybe Text)

The ID of the managed instance targeted by the command. A managed instance can be an EC2 instance or an instance in your hybrid environment that is configured for Systems Manager.

getCommandInvocationResponse_status :: Lens' GetCommandInvocationResponse (Maybe CommandInvocationStatus)

The status of this invocation plugin. This status can be different than StatusDetails.

getCommandInvocationResponse_standardErrorContent :: Lens' GetCommandInvocationResponse (Maybe Text)

The first 8,000 characters written by the plugin to stderr. If the command hasn't finished running, then this string is empty.

getCommandInvocationResponse_cloudWatchOutputConfig :: Lens' GetCommandInvocationResponse (Maybe CloudWatchOutputConfig)

Amazon CloudWatch Logs information where Systems Manager sent the command output.

getCommandInvocationResponse_executionElapsedTime :: Lens' GetCommandInvocationResponse (Maybe Text)

Duration since ExecutionStartDateTime.

getCommandInvocationResponse_documentName :: Lens' GetCommandInvocationResponse (Maybe Text)

The name of the document that was run. For example, AWS-RunShellScript.

getCommandInvocationResponse_standardErrorUrl :: Lens' GetCommandInvocationResponse (Maybe Text)

The URL for the complete text written by the plugin to stderr. If the command hasn't finished running, then this string is empty.

getCommandInvocationResponse_executionStartDateTime :: Lens' GetCommandInvocationResponse (Maybe Text)

The date and time the plugin started running. Date and time are written in ISO 8601 format. For example, June 7, 2017 is represented as 2017-06-7. The following sample Amazon Web Services CLI command uses the InvokedBefore filter.

aws ssm list-commands --filters key=InvokedBefore,value=2017-06-07T00:00:00Z

If the plugin hasn't started to run, the string is empty.

getCommandInvocationResponse_responseCode :: Lens' GetCommandInvocationResponse (Maybe Int)

The error level response code for the plugin script. If the response code is -1, then the command hasn't started running on the instance, or it wasn't received by the instance.

getCommandInvocationResponse_statusDetails :: Lens' GetCommandInvocationResponse (Maybe Text)

A detailed status of the command execution for an invocation. StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Understanding command statuses in the Amazon Web Services Systems Manager User Guide. StatusDetails can be one of the following values:

Pending: The command hasn't been sent to the instance.

In Progress: The command has been sent to the instance but hasn't reached a terminal state.

Delayed: The system attempted to send the command to the target, but the target wasn't available. The instance might not be available because of network issues, because the instance was stopped, or for similar reasons. The system will try to send the command again.

Success: The command or plugin ran successfully. This is a terminal state.

Delivery Timed Out: The command wasn't delivered to the instance before the delivery timeout expired. Delivery timeouts don't count against the parent command's MaxErrors limit, but they do contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Execution Timed Out: The command started to run on the instance, but the execution wasn't complete before the timeout expired. Execution timeouts count against the MaxErrors limit of the parent command. This is a terminal state.

Failed: The command wasn't run successfully on the instance. For a plugin, this indicates that the result code wasn't zero. For a command invocation, this indicates that the result code for one or more plugins wasn't zero. Invocation failures count against the MaxErrors limit of the parent command. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Undeliverable: The command can't be delivered to the instance. The instance might not exist or might not be responding. Undeliverable invocations don't count against the parent command's MaxErrors limit and don't contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Terminated: The parent command exceeded its MaxErrors limit and subsequent command invocations were canceled by the system. This is a terminal state.

getCommandInvocationResponse_executionEndDateTime :: Lens' GetCommandInvocationResponse (Maybe Text)

The date and time the plugin finished running. Date and time are written in ISO 8601 format. For example, June 7, 2017 is represented as 2017-06-7. The following sample Amazon Web Services CLI command uses the InvokedAfter filter.

aws ssm list-commands --filters key=InvokedAfter,value=2017-06-07T00:00:00Z

If the plugin hasn't started to run, the string is empty.

getCommandInvocationResponse_standardOutputUrl :: Lens' GetCommandInvocationResponse (Maybe Text)

The URL for the complete text written by the plugin to stdout in Amazon Simple Storage Service (Amazon S3). If an S3 bucket wasn't specified, then this string is empty.

getCommandInvocationResponse_commandId :: Lens' GetCommandInvocationResponse (Maybe Text)

The parent command ID of the invocation plugin.

getCommandInvocationResponse_documentVersion :: Lens' GetCommandInvocationResponse (Maybe Text)

The Systems Manager document (SSM document) version used in the request.

getCommandInvocationResponse_standardOutputContent :: Lens' GetCommandInvocationResponse (Maybe Text)

The first 24,000 characters written by the plugin to stdout. If the command hasn't finished running, if ExecutionStatus is neither Succeeded nor Failed, then this string is empty.

getCommandInvocationResponse_comment :: Lens' GetCommandInvocationResponse (Maybe Text)

The comment text for the command.

getCommandInvocationResponse_pluginName :: Lens' GetCommandInvocationResponse (Maybe Text)

The name of the plugin, or step name, for which details are reported. For example, aws:RunShellScript is a plugin.

getCommandInvocationResponse_httpStatus :: Lens' GetCommandInvocationResponse Int

The response's http status code.