Network.AWS.SSM.GetConnectionStatus

Description

Retrieves the Session Manager connection status for an instance to determine whether it is running and ready to receive Session Manager connections.

Creating a Request

data GetConnectionStatus

See: newGetConnectionStatus smart constructor.

Constructors

GetConnectionStatus'

Fields

target :: Text

The instance ID.

newGetConnectionStatus

Arguments

:: Text
$sel:target:GetConnectionStatus'

-> GetConnectionStatus

Create a value of GetConnectionStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:target:GetConnectionStatus', getConnectionStatus_target - The instance ID.

Request Lenses

getConnectionStatus_target :: Lens' GetConnectionStatus Text

The instance ID.

Destructuring the Response

data GetConnectionStatusResponse

See: newGetConnectionStatusResponse smart constructor.

Constructors

GetConnectionStatusResponse'

Fields

status :: Maybe ConnectionStatus

The status of the connection to the instance. For example, 'Connected' or 'Not Connected'.

target :: Maybe Text

The ID of the instance to check connection status.

httpStatus :: Int

The response's http status code.

newGetConnectionStatusResponse

Arguments

:: Int
$sel:httpStatus:GetConnectionStatusResponse'

-> GetConnectionStatusResponse

Create a value of GetConnectionStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetConnectionStatusResponse', getConnectionStatusResponse_status - The status of the connection to the instance. For example, 'Connected' or 'Not Connected'.

$sel:target:GetConnectionStatus', getConnectionStatusResponse_target - The ID of the instance to check connection status.

$sel:httpStatus:GetConnectionStatusResponse', getConnectionStatusResponse_httpStatus - The response's http status code.

Response Lenses

getConnectionStatusResponse_status :: Lens' GetConnectionStatusResponse (Maybe ConnectionStatus)

The status of the connection to the instance. For example, 'Connected' or 'Not Connected'.

getConnectionStatusResponse_target :: Lens' GetConnectionStatusResponse (Maybe Text)

The ID of the instance to check connection status.

getConnectionStatusResponse_httpStatus :: Lens' GetConnectionStatusResponse Int

The response's http status code.