Network.AWS.SSM.GetDeployablePatchSnapshotForInstance

Description

Retrieves the current snapshot for the patch baseline the instance uses. This API is primarily used by the AWS-RunPatchBaseline Systems Manager document (SSM document).

If you run the command locally, such as with the Command Line Interface (CLI), the system attempts to use your local Amazon Web Services credentials and the operation fails. To avoid this, you can run the command in the Amazon Web Services Systems Manager console. Use Run Command, a capability of Amazon Web Services Systems Manager, with an SSM document that enables you to target an instance with a script or command. For example, run the command using the AWS-RunShellScript document or the AWS-RunPowerShellScript document.

Creating a Request

data GetDeployablePatchSnapshotForInstance

See: newGetDeployablePatchSnapshotForInstance smart constructor.

Constructors

GetDeployablePatchSnapshotForInstance'

Fields

baselineOverride :: Maybe BaselineOverride

Defines the basic information about a patch baseline override.

instanceId :: Text

The ID of the instance for which the appropriate patch snapshot should be retrieved.

snapshotId :: Text

The snapshot ID provided by the user when running AWS-RunPatchBaseline.

newGetDeployablePatchSnapshotForInstance

Arguments

:: Text
$sel:instanceId:GetDeployablePatchSnapshotForInstance'

-> Text
$sel:snapshotId:GetDeployablePatchSnapshotForInstance'

-> GetDeployablePatchSnapshotForInstance

Create a value of GetDeployablePatchSnapshotForInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baselineOverride:GetDeployablePatchSnapshotForInstance', getDeployablePatchSnapshotForInstance_baselineOverride - Defines the basic information about a patch baseline override.

$sel:instanceId:GetDeployablePatchSnapshotForInstance', getDeployablePatchSnapshotForInstance_instanceId - The ID of the instance for which the appropriate patch snapshot should be retrieved.

$sel:snapshotId:GetDeployablePatchSnapshotForInstance', getDeployablePatchSnapshotForInstance_snapshotId - The snapshot ID provided by the user when running AWS-RunPatchBaseline.

Request Lenses

getDeployablePatchSnapshotForInstance_baselineOverride :: Lens' GetDeployablePatchSnapshotForInstance (Maybe BaselineOverride)

Defines the basic information about a patch baseline override.

getDeployablePatchSnapshotForInstance_instanceId :: Lens' GetDeployablePatchSnapshotForInstance Text

The ID of the instance for which the appropriate patch snapshot should be retrieved.

getDeployablePatchSnapshotForInstance_snapshotId :: Lens' GetDeployablePatchSnapshotForInstance Text

The snapshot ID provided by the user when running AWS-RunPatchBaseline.

Destructuring the Response

data GetDeployablePatchSnapshotForInstanceResponse

See: newGetDeployablePatchSnapshotForInstanceResponse smart constructor.

Constructors

GetDeployablePatchSnapshotForInstanceResponse'

Fields

instanceId :: Maybe Text

The instance ID.

snapshotDownloadUrl :: Maybe Text

A pre-signed Amazon Simple Storage Service (Amazon S3) URL that can be used to download the patch snapshot.

product :: Maybe Text

Returns the specific operating system (for example Windows Server 2012 or Amazon Linux 2015.09) on the instance for the specified patch snapshot.

snapshotId :: Maybe Text

The user-defined snapshot ID.

httpStatus :: Int

The response's http status code.

newGetDeployablePatchSnapshotForInstanceResponse

Arguments

:: Int
$sel:httpStatus:GetDeployablePatchSnapshotForInstanceResponse'

-> GetDeployablePatchSnapshotForInstanceResponse

Create a value of GetDeployablePatchSnapshotForInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:GetDeployablePatchSnapshotForInstance', getDeployablePatchSnapshotForInstanceResponse_instanceId - The instance ID.

$sel:snapshotDownloadUrl:GetDeployablePatchSnapshotForInstanceResponse', getDeployablePatchSnapshotForInstanceResponse_snapshotDownloadUrl - A pre-signed Amazon Simple Storage Service (Amazon S3) URL that can be used to download the patch snapshot.

$sel:product:GetDeployablePatchSnapshotForInstanceResponse', getDeployablePatchSnapshotForInstanceResponse_product - Returns the specific operating system (for example Windows Server 2012 or Amazon Linux 2015.09) on the instance for the specified patch snapshot.

$sel:snapshotId:GetDeployablePatchSnapshotForInstance', getDeployablePatchSnapshotForInstanceResponse_snapshotId - The user-defined snapshot ID.

$sel:httpStatus:GetDeployablePatchSnapshotForInstanceResponse', getDeployablePatchSnapshotForInstanceResponse_httpStatus - The response's http status code.

Response Lenses

getDeployablePatchSnapshotForInstanceResponse_instanceId :: Lens' GetDeployablePatchSnapshotForInstanceResponse (Maybe Text)

The instance ID.

getDeployablePatchSnapshotForInstanceResponse_snapshotDownloadUrl :: Lens' GetDeployablePatchSnapshotForInstanceResponse (Maybe Text)

A pre-signed Amazon Simple Storage Service (Amazon S3) URL that can be used to download the patch snapshot.

getDeployablePatchSnapshotForInstanceResponse_product :: Lens' GetDeployablePatchSnapshotForInstanceResponse (Maybe Text)

Returns the specific operating system (for example Windows Server 2012 or Amazon Linux 2015.09) on the instance for the specified patch snapshot.

getDeployablePatchSnapshotForInstanceResponse_snapshotId :: Lens' GetDeployablePatchSnapshotForInstanceResponse (Maybe Text)

The user-defined snapshot ID.

getDeployablePatchSnapshotForInstanceResponse_httpStatus :: Lens' GetDeployablePatchSnapshotForInstanceResponse Int

The response's http status code.