Network.AWS.SSM.GetMaintenanceWindow

Description

Retrieves a maintenance window.

Creating a Request

data GetMaintenanceWindow

See: newGetMaintenanceWindow smart constructor.

Constructors

GetMaintenanceWindow'

Fields

windowId :: Text

The ID of the maintenance window for which you want to retrieve information.

newGetMaintenanceWindow

Arguments

:: Text
$sel:windowId:GetMaintenanceWindow'

-> GetMaintenanceWindow

Create a value of GetMaintenanceWindow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:windowId:GetMaintenanceWindow', getMaintenanceWindow_windowId - The ID of the maintenance window for which you want to retrieve information.

Request Lenses

getMaintenanceWindow_windowId :: Lens' GetMaintenanceWindow Text

The ID of the maintenance window for which you want to retrieve information.

Destructuring the Response

data GetMaintenanceWindowResponse

See: newGetMaintenanceWindowResponse smart constructor.

Constructors

GetMaintenanceWindowResponse'

Fields

createdDate :: Maybe POSIX

The date the maintenance window was created.

startDate :: Maybe Text

The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become active. The maintenance window won't run before this specified time.

duration :: Maybe Natural

The duration of the maintenance window in hours.

scheduleOffset :: Maybe Natural

The number of days to wait to run a maintenance window after the scheduled cron expression date and time.

enabled :: Maybe Bool

Indicates whether the maintenance window is enabled.

cutoff :: Maybe Natural

The number of hours before the end of the maintenance window that Amazon Web Services Systems Manager stops scheduling new tasks for execution.

modifiedDate :: Maybe POSIX

The date the maintenance window was last modified.

name :: Maybe Text

The name of the maintenance window.

windowId :: Maybe Text

The ID of the created maintenance window.

scheduleTimezone :: Maybe Text

The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format. For example: "America/Los_Angeles", "UTC", or "Asia/Seoul". For more information, see the Time Zone Database on the IANA website.

description :: Maybe (Sensitive Text)

The description of the maintenance window.

endDate :: Maybe Text

The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become inactive. The maintenance window won't run after this specified time.

allowUnassociatedTargets :: Maybe Bool

Whether targets must be registered with the maintenance window before tasks can be defined for those targets.

nextExecutionTime :: Maybe Text

The next time the maintenance window will actually run, taking into account any specified times for the maintenance window to become active or inactive.

schedule :: Maybe Text

The schedule of the maintenance window in the form of a cron or rate expression.

httpStatus :: Int

The response's http status code.

newGetMaintenanceWindowResponse

Arguments

:: Int
$sel:httpStatus:GetMaintenanceWindowResponse'

-> GetMaintenanceWindowResponse

Create a value of GetMaintenanceWindowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdDate:GetMaintenanceWindowResponse', getMaintenanceWindowResponse_createdDate - The date the maintenance window was created.

$sel:startDate:GetMaintenanceWindowResponse', getMaintenanceWindowResponse_startDate - The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become active. The maintenance window won't run before this specified time.

$sel:duration:GetMaintenanceWindowResponse', getMaintenanceWindowResponse_duration - The duration of the maintenance window in hours.

$sel:scheduleOffset:GetMaintenanceWindowResponse', getMaintenanceWindowResponse_scheduleOffset - The number of days to wait to run a maintenance window after the scheduled cron expression date and time.

$sel:enabled:GetMaintenanceWindowResponse', getMaintenanceWindowResponse_enabled - Indicates whether the maintenance window is enabled.

$sel:cutoff:GetMaintenanceWindowResponse', getMaintenanceWindowResponse_cutoff - The number of hours before the end of the maintenance window that Amazon Web Services Systems Manager stops scheduling new tasks for execution.

$sel:modifiedDate:GetMaintenanceWindowResponse', getMaintenanceWindowResponse_modifiedDate - The date the maintenance window was last modified.

$sel:name:GetMaintenanceWindowResponse', getMaintenanceWindowResponse_name - The name of the maintenance window.

$sel:windowId:GetMaintenanceWindow', getMaintenanceWindowResponse_windowId - The ID of the created maintenance window.

$sel:scheduleTimezone:GetMaintenanceWindowResponse', getMaintenanceWindowResponse_scheduleTimezone - The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format. For example: "America/Los_Angeles", "UTC", or "Asia/Seoul". For more information, see the Time Zone Database on the IANA website.

$sel:description:GetMaintenanceWindowResponse', getMaintenanceWindowResponse_description - The description of the maintenance window.

$sel:endDate:GetMaintenanceWindowResponse', getMaintenanceWindowResponse_endDate - The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become inactive. The maintenance window won't run after this specified time.

$sel:allowUnassociatedTargets:GetMaintenanceWindowResponse', getMaintenanceWindowResponse_allowUnassociatedTargets - Whether targets must be registered with the maintenance window before tasks can be defined for those targets.

$sel:nextExecutionTime:GetMaintenanceWindowResponse', getMaintenanceWindowResponse_nextExecutionTime - The next time the maintenance window will actually run, taking into account any specified times for the maintenance window to become active or inactive.

$sel:schedule:GetMaintenanceWindowResponse', getMaintenanceWindowResponse_schedule - The schedule of the maintenance window in the form of a cron or rate expression.

$sel:httpStatus:GetMaintenanceWindowResponse', getMaintenanceWindowResponse_httpStatus - The response's http status code.

Response Lenses

getMaintenanceWindowResponse_createdDate :: Lens' GetMaintenanceWindowResponse (Maybe UTCTime)

The date the maintenance window was created.

getMaintenanceWindowResponse_startDate :: Lens' GetMaintenanceWindowResponse (Maybe Text)

The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become active. The maintenance window won't run before this specified time.

getMaintenanceWindowResponse_duration :: Lens' GetMaintenanceWindowResponse (Maybe Natural)

The duration of the maintenance window in hours.

getMaintenanceWindowResponse_scheduleOffset :: Lens' GetMaintenanceWindowResponse (Maybe Natural)

The number of days to wait to run a maintenance window after the scheduled cron expression date and time.

getMaintenanceWindowResponse_enabled :: Lens' GetMaintenanceWindowResponse (Maybe Bool)

Indicates whether the maintenance window is enabled.

getMaintenanceWindowResponse_cutoff :: Lens' GetMaintenanceWindowResponse (Maybe Natural)

The number of hours before the end of the maintenance window that Amazon Web Services Systems Manager stops scheduling new tasks for execution.

getMaintenanceWindowResponse_modifiedDate :: Lens' GetMaintenanceWindowResponse (Maybe UTCTime)

The date the maintenance window was last modified.

getMaintenanceWindowResponse_name :: Lens' GetMaintenanceWindowResponse (Maybe Text)

The name of the maintenance window.

getMaintenanceWindowResponse_windowId :: Lens' GetMaintenanceWindowResponse (Maybe Text)

The ID of the created maintenance window.

getMaintenanceWindowResponse_scheduleTimezone :: Lens' GetMaintenanceWindowResponse (Maybe Text)

The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format. For example: "America/Los_Angeles", "UTC", or "Asia/Seoul". For more information, see the Time Zone Database on the IANA website.

getMaintenanceWindowResponse_description :: Lens' GetMaintenanceWindowResponse (Maybe Text)

The description of the maintenance window.

getMaintenanceWindowResponse_endDate :: Lens' GetMaintenanceWindowResponse (Maybe Text)

The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become inactive. The maintenance window won't run after this specified time.

getMaintenanceWindowResponse_allowUnassociatedTargets :: Lens' GetMaintenanceWindowResponse (Maybe Bool)

Whether targets must be registered with the maintenance window before tasks can be defined for those targets.

getMaintenanceWindowResponse_nextExecutionTime :: Lens' GetMaintenanceWindowResponse (Maybe Text)

The next time the maintenance window will actually run, taking into account any specified times for the maintenance window to become active or inactive.

getMaintenanceWindowResponse_schedule :: Lens' GetMaintenanceWindowResponse (Maybe Text)

The schedule of the maintenance window in the form of a cron or rate expression.

getMaintenanceWindowResponse_httpStatus :: Lens' GetMaintenanceWindowResponse Int

The response's http status code.