Network.AWS.SSM.GetMaintenanceWindowExecutionTask

Description

Retrieves the details about a specific task run as part of a maintenance window execution.

Creating a Request

data GetMaintenanceWindowExecutionTask

See: newGetMaintenanceWindowExecutionTask smart constructor.

Constructors

GetMaintenanceWindowExecutionTask'

Fields

windowExecutionId :: Text

The ID of the maintenance window execution that includes the task.

taskId :: Text

The ID of the specific task execution in the maintenance window task that should be retrieved.

newGetMaintenanceWindowExecutionTask

Arguments

:: Text
$sel:windowExecutionId:GetMaintenanceWindowExecutionTask'

-> Text
$sel:taskId:GetMaintenanceWindowExecutionTask'

-> GetMaintenanceWindowExecutionTask

Create a value of GetMaintenanceWindowExecutionTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:windowExecutionId:GetMaintenanceWindowExecutionTask', getMaintenanceWindowExecutionTask_windowExecutionId - The ID of the maintenance window execution that includes the task.

$sel:taskId:GetMaintenanceWindowExecutionTask', getMaintenanceWindowExecutionTask_taskId - The ID of the specific task execution in the maintenance window task that should be retrieved.

Request Lenses

getMaintenanceWindowExecutionTask_windowExecutionId :: Lens' GetMaintenanceWindowExecutionTask Text

The ID of the maintenance window execution that includes the task.

getMaintenanceWindowExecutionTask_taskId :: Lens' GetMaintenanceWindowExecutionTask Text

The ID of the specific task execution in the maintenance window task that should be retrieved.

Destructuring the Response

data GetMaintenanceWindowExecutionTaskResponse

See: newGetMaintenanceWindowExecutionTaskResponse smart constructor.

Constructors

GetMaintenanceWindowExecutionTaskResponse'

Fields

maxErrors :: Maybe Text

The defined maximum number of task execution errors allowed before scheduling of the task execution would have been stopped.

taskParameters :: Maybe (Sensitive [Sensitive (HashMap Text (Sensitive MaintenanceWindowTaskParameterValueExpression))])

The parameters passed to the task when it was run.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

The map has the following format:

Key: string, between 1 and 255 characters

Value: an array of strings, each between 1 and 255 characters

status :: Maybe MaintenanceWindowExecutionStatus

The status of the task.

serviceRole :: Maybe Text

The role that was assumed when running the task.

statusDetails :: Maybe Text

The details explaining the status. Not available for all status values.

startTime :: Maybe POSIX

The time the task execution started.

endTime :: Maybe POSIX

The time the task execution completed.

priority :: Maybe Natural

The priority of the task.

maxConcurrency :: Maybe Text

The defined maximum number of task executions that could be run in parallel.

windowExecutionId :: Maybe Text

The ID of the maintenance window execution that includes the task.

taskArn :: Maybe Text

The Amazon Resource Name (ARN) of the task that ran.

type' :: Maybe MaintenanceWindowTaskType

The type of task that was run.

taskExecutionId :: Maybe Text

The ID of the specific task execution in the maintenance window task that was retrieved.

httpStatus :: Int

The response's http status code.

newGetMaintenanceWindowExecutionTaskResponse

Arguments

:: Int
$sel:httpStatus:GetMaintenanceWindowExecutionTaskResponse'

-> GetMaintenanceWindowExecutionTaskResponse

Create a value of GetMaintenanceWindowExecutionTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxErrors:GetMaintenanceWindowExecutionTaskResponse', getMaintenanceWindowExecutionTaskResponse_maxErrors - The defined maximum number of task execution errors allowed before scheduling of the task execution would have been stopped.

$sel:taskParameters:GetMaintenanceWindowExecutionTaskResponse', getMaintenanceWindowExecutionTaskResponse_taskParameters - The parameters passed to the task when it was run.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

The map has the following format:

Key: string, between 1 and 255 characters

Value: an array of strings, each between 1 and 255 characters

$sel:status:GetMaintenanceWindowExecutionTaskResponse', getMaintenanceWindowExecutionTaskResponse_status - The status of the task.

$sel:serviceRole:GetMaintenanceWindowExecutionTaskResponse', getMaintenanceWindowExecutionTaskResponse_serviceRole - The role that was assumed when running the task.

$sel:statusDetails:GetMaintenanceWindowExecutionTaskResponse', getMaintenanceWindowExecutionTaskResponse_statusDetails - The details explaining the status. Not available for all status values.

$sel:startTime:GetMaintenanceWindowExecutionTaskResponse', getMaintenanceWindowExecutionTaskResponse_startTime - The time the task execution started.

$sel:endTime:GetMaintenanceWindowExecutionTaskResponse', getMaintenanceWindowExecutionTaskResponse_endTime - The time the task execution completed.

$sel:priority:GetMaintenanceWindowExecutionTaskResponse', getMaintenanceWindowExecutionTaskResponse_priority - The priority of the task.

$sel:maxConcurrency:GetMaintenanceWindowExecutionTaskResponse', getMaintenanceWindowExecutionTaskResponse_maxConcurrency - The defined maximum number of task executions that could be run in parallel.

$sel:windowExecutionId:GetMaintenanceWindowExecutionTask', getMaintenanceWindowExecutionTaskResponse_windowExecutionId - The ID of the maintenance window execution that includes the task.

$sel:taskArn:GetMaintenanceWindowExecutionTaskResponse', getMaintenanceWindowExecutionTaskResponse_taskArn - The Amazon Resource Name (ARN) of the task that ran.

$sel:type':GetMaintenanceWindowExecutionTaskResponse', getMaintenanceWindowExecutionTaskResponse_type - The type of task that was run.

$sel:taskExecutionId:GetMaintenanceWindowExecutionTaskResponse', getMaintenanceWindowExecutionTaskResponse_taskExecutionId - The ID of the specific task execution in the maintenance window task that was retrieved.

$sel:httpStatus:GetMaintenanceWindowExecutionTaskResponse', getMaintenanceWindowExecutionTaskResponse_httpStatus - The response's http status code.

Response Lenses

getMaintenanceWindowExecutionTaskResponse_maxErrors :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe Text)

The defined maximum number of task execution errors allowed before scheduling of the task execution would have been stopped.

getMaintenanceWindowExecutionTaskResponse_taskParameters :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe [HashMap Text MaintenanceWindowTaskParameterValueExpression])

The parameters passed to the task when it was run.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

The map has the following format:

Key: string, between 1 and 255 characters

Value: an array of strings, each between 1 and 255 characters

getMaintenanceWindowExecutionTaskResponse_status :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe MaintenanceWindowExecutionStatus)

The status of the task.

getMaintenanceWindowExecutionTaskResponse_serviceRole :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe Text)

The role that was assumed when running the task.

getMaintenanceWindowExecutionTaskResponse_statusDetails :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe Text)

The details explaining the status. Not available for all status values.

getMaintenanceWindowExecutionTaskResponse_startTime :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe UTCTime)

The time the task execution started.

getMaintenanceWindowExecutionTaskResponse_endTime :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe UTCTime)

The time the task execution completed.

getMaintenanceWindowExecutionTaskResponse_priority :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe Natural)

The priority of the task.

getMaintenanceWindowExecutionTaskResponse_maxConcurrency :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe Text)

The defined maximum number of task executions that could be run in parallel.

getMaintenanceWindowExecutionTaskResponse_windowExecutionId :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe Text)

The ID of the maintenance window execution that includes the task.

getMaintenanceWindowExecutionTaskResponse_taskArn :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe Text)

The Amazon Resource Name (ARN) of the task that ran.

getMaintenanceWindowExecutionTaskResponse_type :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe MaintenanceWindowTaskType)

The type of task that was run.

getMaintenanceWindowExecutionTaskResponse_taskExecutionId :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe Text)

The ID of the specific task execution in the maintenance window task that was retrieved.

getMaintenanceWindowExecutionTaskResponse_httpStatus :: Lens' GetMaintenanceWindowExecutionTaskResponse Int

The response's http status code.