Network.AWS.SSM.GetMaintenanceWindowTask

Description

Lists the tasks in a maintenance window.

For maintenance window tasks without a specified target, you can't supply values for --max-errors and --max-concurrency. Instead, the system inserts a placeholder value of 1, which may be reported in the response to this command. These values don't affect the running of your task and can be ignored.

Creating a Request

data GetMaintenanceWindowTask

See: newGetMaintenanceWindowTask smart constructor.

Constructors

GetMaintenanceWindowTask'

Fields

windowId :: Text

The maintenance window ID that includes the task to retrieve.

windowTaskId :: Text

The maintenance window task ID to retrieve.

newGetMaintenanceWindowTask

Arguments

:: Text
$sel:windowId:GetMaintenanceWindowTask'

-> Text
$sel:windowTaskId:GetMaintenanceWindowTask'

-> GetMaintenanceWindowTask

Create a value of GetMaintenanceWindowTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:windowId:GetMaintenanceWindowTask', getMaintenanceWindowTask_windowId - The maintenance window ID that includes the task to retrieve.

$sel:windowTaskId:GetMaintenanceWindowTask', getMaintenanceWindowTask_windowTaskId - The maintenance window task ID to retrieve.

Request Lenses

getMaintenanceWindowTask_windowId :: Lens' GetMaintenanceWindowTask Text

The maintenance window ID that includes the task to retrieve.

getMaintenanceWindowTask_windowTaskId :: Lens' GetMaintenanceWindowTask Text

The maintenance window task ID to retrieve.

Destructuring the Response

data GetMaintenanceWindowTaskResponse

See: newGetMaintenanceWindowTaskResponse smart constructor.

Constructors

GetMaintenanceWindowTaskResponse'

Fields

serviceRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) service role to use to publish Amazon Simple Notification Service (Amazon SNS) notifications for maintenance window Run Command tasks.

windowTaskId :: Maybe Text

The retrieved maintenance window task ID.

taskParameters :: Maybe (Sensitive (HashMap Text (Sensitive MaintenanceWindowTaskParameterValueExpression)))

The parameters to pass to the task when it runs.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

priority :: Maybe Natural

The priority of the task when it runs. The lower the number, the higher the priority. Tasks that have the same priority are scheduled in parallel.

taskType :: Maybe MaintenanceWindowTaskType

The type of task to run.

taskArn :: Maybe Text

The resource that the task used during execution. For RUN_COMMAND and AUTOMATION task types, the value of TaskArn is the SSM document name/ARN. For LAMBDA tasks, the value is the function name/ARN. For STEP_FUNCTIONS tasks, the value is the state machine ARN.

cutoffBehavior :: Maybe MaintenanceWindowTaskCutoffBehavior

The action to take on tasks when the maintenance window cutoff time is reached. CONTINUE_TASK means that tasks continue to run. For Automation, Lambda, Step Functions tasks, CANCEL_TASK means that currently running task invocations continue, but no new task invocations are started. For Run Command tasks, CANCEL_TASK means the system attempts to stop the task by sending a CancelCommand operation.

maxErrors :: Maybe Text

The maximum number of errors allowed before the task stops being scheduled.

For maintenance window tasks without a target specified, you can't supply a value for this option. Instead, the system inserts a placeholder value of 1, which may be reported in the response to this command. This value doesn't affect the running of your task and can be ignored.

taskInvocationParameters :: Maybe MaintenanceWindowTaskInvocationParameters

The parameters to pass to the task when it runs.

name :: Maybe Text

The retrieved task name.

targets :: Maybe [Target]

The targets where the task should run.

loggingInfo :: Maybe LoggingInfo

The location in Amazon Simple Storage Service (Amazon S3) where the task results are logged.

LoggingInfo has been deprecated. To specify an Amazon Simple Storage Service (Amazon S3) bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Amazon Web Services Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

description :: Maybe (Sensitive Text)

The retrieved task description.

maxConcurrency :: Maybe Text

The maximum number of targets allowed to run this task in parallel.

For maintenance window tasks without a target specified, you can't supply a value for this option. Instead, the system inserts a placeholder value of 1, which may be reported in the response to this command. This value doesn't affect the running of your task and can be ignored.

windowId :: Maybe Text

The retrieved maintenance window ID.

httpStatus :: Int

The response's http status code.

newGetMaintenanceWindowTaskResponse

Arguments

:: Int
$sel:httpStatus:GetMaintenanceWindowTaskResponse'

-> GetMaintenanceWindowTaskResponse

Create a value of GetMaintenanceWindowTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceRoleArn:GetMaintenanceWindowTaskResponse', getMaintenanceWindowTaskResponse_serviceRoleArn - The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) service role to use to publish Amazon Simple Notification Service (Amazon SNS) notifications for maintenance window Run Command tasks.

$sel:windowTaskId:GetMaintenanceWindowTask', getMaintenanceWindowTaskResponse_windowTaskId - The retrieved maintenance window task ID.

$sel:taskParameters:GetMaintenanceWindowTaskResponse', getMaintenanceWindowTaskResponse_taskParameters - The parameters to pass to the task when it runs.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

$sel:priority:GetMaintenanceWindowTaskResponse', getMaintenanceWindowTaskResponse_priority - The priority of the task when it runs. The lower the number, the higher the priority. Tasks that have the same priority are scheduled in parallel.

$sel:taskType:GetMaintenanceWindowTaskResponse', getMaintenanceWindowTaskResponse_taskType - The type of task to run.

$sel:taskArn:GetMaintenanceWindowTaskResponse', getMaintenanceWindowTaskResponse_taskArn - The resource that the task used during execution. For RUN_COMMAND and AUTOMATION task types, the value of TaskArn is the SSM document name/ARN. For LAMBDA tasks, the value is the function name/ARN. For STEP_FUNCTIONS tasks, the value is the state machine ARN.

$sel:cutoffBehavior:GetMaintenanceWindowTaskResponse', getMaintenanceWindowTaskResponse_cutoffBehavior - The action to take on tasks when the maintenance window cutoff time is reached. CONTINUE_TASK means that tasks continue to run. For Automation, Lambda, Step Functions tasks, CANCEL_TASK means that currently running task invocations continue, but no new task invocations are started. For Run Command tasks, CANCEL_TASK means the system attempts to stop the task by sending a CancelCommand operation.

$sel:maxErrors:GetMaintenanceWindowTaskResponse', getMaintenanceWindowTaskResponse_maxErrors - The maximum number of errors allowed before the task stops being scheduled.

For maintenance window tasks without a target specified, you can't supply a value for this option. Instead, the system inserts a placeholder value of 1, which may be reported in the response to this command. This value doesn't affect the running of your task and can be ignored.

$sel:taskInvocationParameters:GetMaintenanceWindowTaskResponse', getMaintenanceWindowTaskResponse_taskInvocationParameters - The parameters to pass to the task when it runs.

$sel:name:GetMaintenanceWindowTaskResponse', getMaintenanceWindowTaskResponse_name - The retrieved task name.

$sel:targets:GetMaintenanceWindowTaskResponse', getMaintenanceWindowTaskResponse_targets - The targets where the task should run.

$sel:loggingInfo:GetMaintenanceWindowTaskResponse', getMaintenanceWindowTaskResponse_loggingInfo - The location in Amazon Simple Storage Service (Amazon S3) where the task results are logged.

LoggingInfo has been deprecated. To specify an Amazon Simple Storage Service (Amazon S3) bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Amazon Web Services Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

$sel:description:GetMaintenanceWindowTaskResponse', getMaintenanceWindowTaskResponse_description - The retrieved task description.

$sel:maxConcurrency:GetMaintenanceWindowTaskResponse', getMaintenanceWindowTaskResponse_maxConcurrency - The maximum number of targets allowed to run this task in parallel.

For maintenance window tasks without a target specified, you can't supply a value for this option. Instead, the system inserts a placeholder value of 1, which may be reported in the response to this command. This value doesn't affect the running of your task and can be ignored.

$sel:windowId:GetMaintenanceWindowTask', getMaintenanceWindowTaskResponse_windowId - The retrieved maintenance window ID.

$sel:httpStatus:GetMaintenanceWindowTaskResponse', getMaintenanceWindowTaskResponse_httpStatus - The response's http status code.

Response Lenses

getMaintenanceWindowTaskResponse_serviceRoleArn :: Lens' GetMaintenanceWindowTaskResponse (Maybe Text)

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) service role to use to publish Amazon Simple Notification Service (Amazon SNS) notifications for maintenance window Run Command tasks.

getMaintenanceWindowTaskResponse_windowTaskId :: Lens' GetMaintenanceWindowTaskResponse (Maybe Text)

The retrieved maintenance window task ID.

getMaintenanceWindowTaskResponse_taskParameters :: Lens' GetMaintenanceWindowTaskResponse (Maybe (HashMap Text MaintenanceWindowTaskParameterValueExpression))

The parameters to pass to the task when it runs.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

getMaintenanceWindowTaskResponse_priority :: Lens' GetMaintenanceWindowTaskResponse (Maybe Natural)

The priority of the task when it runs. The lower the number, the higher the priority. Tasks that have the same priority are scheduled in parallel.

getMaintenanceWindowTaskResponse_taskType :: Lens' GetMaintenanceWindowTaskResponse (Maybe MaintenanceWindowTaskType)

The type of task to run.

getMaintenanceWindowTaskResponse_taskArn :: Lens' GetMaintenanceWindowTaskResponse (Maybe Text)

The resource that the task used during execution. For RUN_COMMAND and AUTOMATION task types, the value of TaskArn is the SSM document name/ARN. For LAMBDA tasks, the value is the function name/ARN. For STEP_FUNCTIONS tasks, the value is the state machine ARN.

getMaintenanceWindowTaskResponse_cutoffBehavior :: Lens' GetMaintenanceWindowTaskResponse (Maybe MaintenanceWindowTaskCutoffBehavior)

The action to take on tasks when the maintenance window cutoff time is reached. CONTINUE_TASK means that tasks continue to run. For Automation, Lambda, Step Functions tasks, CANCEL_TASK means that currently running task invocations continue, but no new task invocations are started. For Run Command tasks, CANCEL_TASK means the system attempts to stop the task by sending a CancelCommand operation.

getMaintenanceWindowTaskResponse_maxErrors :: Lens' GetMaintenanceWindowTaskResponse (Maybe Text)

The maximum number of errors allowed before the task stops being scheduled.

For maintenance window tasks without a target specified, you can't supply a value for this option. Instead, the system inserts a placeholder value of 1, which may be reported in the response to this command. This value doesn't affect the running of your task and can be ignored.

getMaintenanceWindowTaskResponse_taskInvocationParameters :: Lens' GetMaintenanceWindowTaskResponse (Maybe MaintenanceWindowTaskInvocationParameters)

The parameters to pass to the task when it runs.

getMaintenanceWindowTaskResponse_name :: Lens' GetMaintenanceWindowTaskResponse (Maybe Text)

The retrieved task name.

getMaintenanceWindowTaskResponse_targets :: Lens' GetMaintenanceWindowTaskResponse (Maybe [Target])

The targets where the task should run.

getMaintenanceWindowTaskResponse_loggingInfo :: Lens' GetMaintenanceWindowTaskResponse (Maybe LoggingInfo)

The location in Amazon Simple Storage Service (Amazon S3) where the task results are logged.

LoggingInfo has been deprecated. To specify an Amazon Simple Storage Service (Amazon S3) bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Amazon Web Services Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

getMaintenanceWindowTaskResponse_description :: Lens' GetMaintenanceWindowTaskResponse (Maybe Text)

The retrieved task description.

getMaintenanceWindowTaskResponse_maxConcurrency :: Lens' GetMaintenanceWindowTaskResponse (Maybe Text)

The maximum number of targets allowed to run this task in parallel.

For maintenance window tasks without a target specified, you can't supply a value for this option. Instead, the system inserts a placeholder value of 1, which may be reported in the response to this command. This value doesn't affect the running of your task and can be ignored.

getMaintenanceWindowTaskResponse_windowId :: Lens' GetMaintenanceWindowTaskResponse (Maybe Text)

The retrieved maintenance window ID.

getMaintenanceWindowTaskResponse_httpStatus :: Lens' GetMaintenanceWindowTaskResponse Int

The response's http status code.