Network.AWS.SSM.GetParametersByPath

Description

Retrieve information about one or more parameters in a specific hierarchy.

Request results are returned on a best-effort basis. If you specify MaxResults in the request, the response includes information up to the limit specified. The number of items returned, however, can be between zero and the value of MaxResults. If the service reaches an internal limit while processing the results, it stops the operation and returns the matching values up to that point and a NextToken. You can specify the NextToken in a subsequent call to get the next set of results.

This operation returns paginated results.

Creating a Request

data GetParametersByPath

See: newGetParametersByPath smart constructor.

Constructors

GetParametersByPath'

Fields

withDecryption :: Maybe Bool

Retrieve all parameters in a hierarchy with their value decrypted.

nextToken :: Maybe Text

A token to start the list. Use this token to get the next set of results.

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

parameterFilters :: Maybe [ParameterStringFilter]

Filters to limit the request results.

The following Key values are supported for GetParametersByPath: Type, KeyId, and Label.

The following Key values aren't supported for GetParametersByPath: tag, DataType, Name, Path, and Tier.

recursive :: Maybe Bool

Retrieve all parameters within a hierarchy.

If a user has access to a path, then the user can access all levels of that path. For example, if a user has permission to access path /a, then the user can also access /a/b. Even if a user has explicitly been denied access in IAM for parameter /a/b, they can still call the GetParametersByPath API operation recursively for /a and view /a/b.

path :: Text

The hierarchy for the parameter. Hierarchies start with a forward slash (/). The hierachy is the parameter name except the last part of the parameter. For the API call to succeeed, the last part of the parameter name can't be in the path. A parameter name hierarchy can have a maximum of 15 levels. Here is an example of a hierarchy: /Finance/Prod/IAD/WinServ2016/license33

newGetParametersByPath

Arguments

:: Text
$sel:path:GetParametersByPath'

-> GetParametersByPath

Create a value of GetParametersByPath with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:withDecryption:GetParametersByPath', getParametersByPath_withDecryption - Retrieve all parameters in a hierarchy with their value decrypted.

$sel:nextToken:GetParametersByPath', getParametersByPath_nextToken - A token to start the list. Use this token to get the next set of results.

$sel:maxResults:GetParametersByPath', getParametersByPath_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

$sel:parameterFilters:GetParametersByPath', getParametersByPath_parameterFilters - Filters to limit the request results.

The following Key values are supported for GetParametersByPath: Type, KeyId, and Label.

The following Key values aren't supported for GetParametersByPath: tag, DataType, Name, Path, and Tier.

$sel:recursive:GetParametersByPath', getParametersByPath_recursive - Retrieve all parameters within a hierarchy.

If a user has access to a path, then the user can access all levels of that path. For example, if a user has permission to access path /a, then the user can also access /a/b. Even if a user has explicitly been denied access in IAM for parameter /a/b, they can still call the GetParametersByPath API operation recursively for /a and view /a/b.

$sel:path:GetParametersByPath', getParametersByPath_path - The hierarchy for the parameter. Hierarchies start with a forward slash (/). The hierachy is the parameter name except the last part of the parameter. For the API call to succeeed, the last part of the parameter name can't be in the path. A parameter name hierarchy can have a maximum of 15 levels. Here is an example of a hierarchy: /Finance/Prod/IAD/WinServ2016/license33

Request Lenses

getParametersByPath_withDecryption :: Lens' GetParametersByPath (Maybe Bool)

Retrieve all parameters in a hierarchy with their value decrypted.

getParametersByPath_nextToken :: Lens' GetParametersByPath (Maybe Text)

A token to start the list. Use this token to get the next set of results.

getParametersByPath_maxResults :: Lens' GetParametersByPath (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

getParametersByPath_parameterFilters :: Lens' GetParametersByPath (Maybe [ParameterStringFilter])

Filters to limit the request results.

The following Key values are supported for GetParametersByPath: Type, KeyId, and Label.

The following Key values aren't supported for GetParametersByPath: tag, DataType, Name, Path, and Tier.

getParametersByPath_recursive :: Lens' GetParametersByPath (Maybe Bool)

Retrieve all parameters within a hierarchy.

If a user has access to a path, then the user can access all levels of that path. For example, if a user has permission to access path /a, then the user can also access /a/b. Even if a user has explicitly been denied access in IAM for parameter /a/b, they can still call the GetParametersByPath API operation recursively for /a and view /a/b.

getParametersByPath_path :: Lens' GetParametersByPath Text

The hierarchy for the parameter. Hierarchies start with a forward slash (/). The hierachy is the parameter name except the last part of the parameter. For the API call to succeeed, the last part of the parameter name can't be in the path. A parameter name hierarchy can have a maximum of 15 levels. Here is an example of a hierarchy: /Finance/Prod/IAD/WinServ2016/license33

Destructuring the Response

data GetParametersByPathResponse

See: newGetParametersByPathResponse smart constructor.

Constructors

GetParametersByPathResponse'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. Use this token to get the next set of results.

parameters :: Maybe [Parameter]

A list of parameters found in the specified hierarchy.

httpStatus :: Int

The response's http status code.

newGetParametersByPathResponse

Arguments

:: Int
$sel:httpStatus:GetParametersByPathResponse'

-> GetParametersByPathResponse

Create a value of GetParametersByPathResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetParametersByPath', getParametersByPathResponse_nextToken - The token for the next set of items to return. Use this token to get the next set of results.

$sel:parameters:GetParametersByPathResponse', getParametersByPathResponse_parameters - A list of parameters found in the specified hierarchy.

$sel:httpStatus:GetParametersByPathResponse', getParametersByPathResponse_httpStatus - The response's http status code.

Response Lenses

getParametersByPathResponse_nextToken :: Lens' GetParametersByPathResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

getParametersByPathResponse_parameters :: Lens' GetParametersByPathResponse (Maybe [Parameter])

A list of parameters found in the specified hierarchy.

getParametersByPathResponse_httpStatus :: Lens' GetParametersByPathResponse Int

The response's http status code.