Network.AWS.SSM.GetPatchBaseline

Description

Retrieves information about a patch baseline.

Creating a Request

data GetPatchBaseline

See: newGetPatchBaseline smart constructor.

Constructors

GetPatchBaseline'

Fields

baselineId :: Text

The ID of the patch baseline to retrieve.

To retrieve information about an Amazon Web Services managed patch baseline, specify the full Amazon Resource Name (ARN) of the baseline. For example, for the baseline AWS-AmazonLinuxDefaultPatchBaseline, specify arn:aws:ssm:us-east-2:733109147000:patchbaseline/pb-0e392de35e7c563b7 instead of pb-0e392de35e7c563b7.

newGetPatchBaseline

Arguments

:: Text
$sel:baselineId:GetPatchBaseline'

-> GetPatchBaseline

Create a value of GetPatchBaseline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baselineId:GetPatchBaseline', getPatchBaseline_baselineId - The ID of the patch baseline to retrieve.

To retrieve information about an Amazon Web Services managed patch baseline, specify the full Amazon Resource Name (ARN) of the baseline. For example, for the baseline AWS-AmazonLinuxDefaultPatchBaseline, specify arn:aws:ssm:us-east-2:733109147000:patchbaseline/pb-0e392de35e7c563b7 instead of pb-0e392de35e7c563b7.

Request Lenses

getPatchBaseline_baselineId :: Lens' GetPatchBaseline Text

The ID of the patch baseline to retrieve.

To retrieve information about an Amazon Web Services managed patch baseline, specify the full Amazon Resource Name (ARN) of the baseline. For example, for the baseline AWS-AmazonLinuxDefaultPatchBaseline, specify arn:aws:ssm:us-east-2:733109147000:patchbaseline/pb-0e392de35e7c563b7 instead of pb-0e392de35e7c563b7.

Destructuring the Response

data GetPatchBaselineResponse

See: newGetPatchBaselineResponse smart constructor.

Constructors

GetPatchBaselineResponse'

Fields

createdDate :: Maybe POSIX

The date the patch baseline was created.

rejectedPatches :: Maybe [Text]

A list of explicitly rejected patches for the baseline.

baselineId :: Maybe Text

The ID of the retrieved patch baseline.

sources :: Maybe [PatchSource]

Information about the patches to use to update the instances, including target operating systems and source repositories. Applies to Linux instances only.

approvedPatchesEnableNonSecurity :: Maybe Bool

Indicates whether the list of approved patches includes non-security updates that should be applied to the instances. The default value is false. Applies to Linux instances only.

approvedPatchesComplianceLevel :: Maybe PatchComplianceLevel

Returns the specified compliance severity level for approved patches in the patch baseline.

modifiedDate :: Maybe POSIX

The date the patch baseline was last modified.

patchGroups :: Maybe [Text]

Patch groups included in the patch baseline.

name :: Maybe Text

The name of the patch baseline.

description :: Maybe Text

A description of the patch baseline.

approvedPatches :: Maybe [Text]

A list of explicitly approved patches for the baseline.

rejectedPatchesAction :: Maybe PatchAction

The action specified to take on patches included in the RejectedPatches list. A patch can be allowed only if it is a dependency of another package, or blocked entirely along with packages that include it as a dependency.

operatingSystem :: Maybe OperatingSystem

Returns the operating system specified for the patch baseline.

globalFilters :: Maybe PatchFilterGroup

A set of global filters used to exclude patches from the baseline.

approvalRules :: Maybe PatchRuleGroup

A set of rules used to include patches in the baseline.

httpStatus :: Int

The response's http status code.

newGetPatchBaselineResponse

Arguments

:: Int
$sel:httpStatus:GetPatchBaselineResponse'

-> GetPatchBaselineResponse

Create a value of GetPatchBaselineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdDate:GetPatchBaselineResponse', getPatchBaselineResponse_createdDate - The date the patch baseline was created.

$sel:rejectedPatches:GetPatchBaselineResponse', getPatchBaselineResponse_rejectedPatches - A list of explicitly rejected patches for the baseline.

$sel:baselineId:GetPatchBaseline', getPatchBaselineResponse_baselineId - The ID of the retrieved patch baseline.

$sel:sources:GetPatchBaselineResponse', getPatchBaselineResponse_sources - Information about the patches to use to update the instances, including target operating systems and source repositories. Applies to Linux instances only.

$sel:approvedPatchesEnableNonSecurity:GetPatchBaselineResponse', getPatchBaselineResponse_approvedPatchesEnableNonSecurity - Indicates whether the list of approved patches includes non-security updates that should be applied to the instances. The default value is false. Applies to Linux instances only.

$sel:approvedPatchesComplianceLevel:GetPatchBaselineResponse', getPatchBaselineResponse_approvedPatchesComplianceLevel - Returns the specified compliance severity level for approved patches in the patch baseline.

$sel:modifiedDate:GetPatchBaselineResponse', getPatchBaselineResponse_modifiedDate - The date the patch baseline was last modified.

$sel:patchGroups:GetPatchBaselineResponse', getPatchBaselineResponse_patchGroups - Patch groups included in the patch baseline.

$sel:name:GetPatchBaselineResponse', getPatchBaselineResponse_name - The name of the patch baseline.

$sel:description:GetPatchBaselineResponse', getPatchBaselineResponse_description - A description of the patch baseline.

$sel:approvedPatches:GetPatchBaselineResponse', getPatchBaselineResponse_approvedPatches - A list of explicitly approved patches for the baseline.

$sel:rejectedPatchesAction:GetPatchBaselineResponse', getPatchBaselineResponse_rejectedPatchesAction - The action specified to take on patches included in the RejectedPatches list. A patch can be allowed only if it is a dependency of another package, or blocked entirely along with packages that include it as a dependency.

$sel:operatingSystem:GetPatchBaselineResponse', getPatchBaselineResponse_operatingSystem - Returns the operating system specified for the patch baseline.

$sel:globalFilters:GetPatchBaselineResponse', getPatchBaselineResponse_globalFilters - A set of global filters used to exclude patches from the baseline.

$sel:approvalRules:GetPatchBaselineResponse', getPatchBaselineResponse_approvalRules - A set of rules used to include patches in the baseline.

$sel:httpStatus:GetPatchBaselineResponse', getPatchBaselineResponse_httpStatus - The response's http status code.

Response Lenses

getPatchBaselineResponse_createdDate :: Lens' GetPatchBaselineResponse (Maybe UTCTime)

The date the patch baseline was created.

getPatchBaselineResponse_rejectedPatches :: Lens' GetPatchBaselineResponse (Maybe [Text])

A list of explicitly rejected patches for the baseline.

getPatchBaselineResponse_baselineId :: Lens' GetPatchBaselineResponse (Maybe Text)

The ID of the retrieved patch baseline.

getPatchBaselineResponse_sources :: Lens' GetPatchBaselineResponse (Maybe [PatchSource])

Information about the patches to use to update the instances, including target operating systems and source repositories. Applies to Linux instances only.

getPatchBaselineResponse_approvedPatchesEnableNonSecurity :: Lens' GetPatchBaselineResponse (Maybe Bool)

Indicates whether the list of approved patches includes non-security updates that should be applied to the instances. The default value is false. Applies to Linux instances only.

getPatchBaselineResponse_approvedPatchesComplianceLevel :: Lens' GetPatchBaselineResponse (Maybe PatchComplianceLevel)

Returns the specified compliance severity level for approved patches in the patch baseline.

getPatchBaselineResponse_modifiedDate :: Lens' GetPatchBaselineResponse (Maybe UTCTime)

The date the patch baseline was last modified.

getPatchBaselineResponse_patchGroups :: Lens' GetPatchBaselineResponse (Maybe [Text])

Patch groups included in the patch baseline.

getPatchBaselineResponse_name :: Lens' GetPatchBaselineResponse (Maybe Text)

The name of the patch baseline.

getPatchBaselineResponse_description :: Lens' GetPatchBaselineResponse (Maybe Text)

A description of the patch baseline.

getPatchBaselineResponse_approvedPatches :: Lens' GetPatchBaselineResponse (Maybe [Text])

A list of explicitly approved patches for the baseline.

getPatchBaselineResponse_rejectedPatchesAction :: Lens' GetPatchBaselineResponse (Maybe PatchAction)

The action specified to take on patches included in the RejectedPatches list. A patch can be allowed only if it is a dependency of another package, or blocked entirely along with packages that include it as a dependency.

getPatchBaselineResponse_operatingSystem :: Lens' GetPatchBaselineResponse (Maybe OperatingSystem)

Returns the operating system specified for the patch baseline.

getPatchBaselineResponse_globalFilters :: Lens' GetPatchBaselineResponse (Maybe PatchFilterGroup)

A set of global filters used to exclude patches from the baseline.

getPatchBaselineResponse_approvalRules :: Lens' GetPatchBaselineResponse (Maybe PatchRuleGroup)

A set of rules used to include patches in the baseline.

getPatchBaselineResponse_httpStatus :: Lens' GetPatchBaselineResponse Int

The response's http status code.