Network.AWS.SSM.GetServiceSetting

Description

ServiceSetting is an account-level setting for an Amazon Web Services service. This setting defines how a user interacts with or uses a service or a feature of a service. For example, if an Amazon Web Services service charges money to the account based on feature or service usage, then the Amazon Web Services service team might create a default setting of false. This means the user can't use this feature unless they change the setting to true and intentionally opt in for a paid feature.

Services map a SettingId object to a setting value. Amazon Web Services services teams define the default value for a SettingId. You can't create a new SettingId, but you can overwrite the default value if you have the ssm:UpdateServiceSetting permission for the setting. Use the UpdateServiceSetting API operation to change the default setting. Or use the ResetServiceSetting to change the value back to the original value defined by the Amazon Web Services service team.

Query the current service setting for the Amazon Web Services account.

Creating a Request

data GetServiceSetting

The request body of the GetServiceSetting API operation.

See: newGetServiceSetting smart constructor.

Constructors

GetServiceSetting'

Fields

settingId :: Text

The ID of the service setting to get. The setting ID can be one of the following.

/ssm/automation/customer-script-log-destination

/ssm/automation/customer-script-log-group-name

/ssm/documents/console/public-sharing-permission

/ssm/parameter-store/default-parameter-tier

/ssm/parameter-store/high-throughput-enabled

/ssm/managed-instance/activation-tier

newGetServiceSetting

Arguments

:: Text
$sel:settingId:GetServiceSetting'

-> GetServiceSetting

Create a value of GetServiceSetting with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:settingId:GetServiceSetting', getServiceSetting_settingId - The ID of the service setting to get. The setting ID can be one of the following.

/ssm/automation/customer-script-log-destination

/ssm/automation/customer-script-log-group-name

/ssm/documents/console/public-sharing-permission

/ssm/parameter-store/default-parameter-tier

/ssm/parameter-store/high-throughput-enabled

/ssm/managed-instance/activation-tier

Request Lenses

getServiceSetting_settingId :: Lens' GetServiceSetting Text

The ID of the service setting to get. The setting ID can be one of the following.

/ssm/automation/customer-script-log-destination

/ssm/automation/customer-script-log-group-name

/ssm/documents/console/public-sharing-permission

/ssm/parameter-store/default-parameter-tier

/ssm/parameter-store/high-throughput-enabled

/ssm/managed-instance/activation-tier

Destructuring the Response

data GetServiceSettingResponse

The query result body of the GetServiceSetting API operation.

See: newGetServiceSettingResponse smart constructor.

Constructors

GetServiceSettingResponse'

Fields

serviceSetting :: Maybe ServiceSetting

The query result of the current service setting.

httpStatus :: Int

The response's http status code.

newGetServiceSettingResponse

Arguments

:: Int
$sel:httpStatus:GetServiceSettingResponse'

-> GetServiceSettingResponse

Create a value of GetServiceSettingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceSetting:GetServiceSettingResponse', getServiceSettingResponse_serviceSetting - The query result of the current service setting.

$sel:httpStatus:GetServiceSettingResponse', getServiceSettingResponse_httpStatus - The response's http status code.

Response Lenses

getServiceSettingResponse_serviceSetting :: Lens' GetServiceSettingResponse (Maybe ServiceSetting)

The query result of the current service setting.

getServiceSettingResponse_httpStatus :: Lens' GetServiceSettingResponse Int

The response's http status code.