Network.AWS.SSM.ListAssociationVersions

Description

Retrieves all versions of an association for a specific association ID.

This operation returns paginated results.

Creating a Request

data ListAssociationVersions

See: newListAssociationVersions smart constructor.

Constructors

ListAssociationVersions'

Fields

nextToken :: Maybe Text

A token to start the list. Use this token to get the next set of results.

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

associationId :: Text

The association ID for which you want to view all versions.

newListAssociationVersions

Arguments

:: Text
$sel:associationId:ListAssociationVersions'

-> ListAssociationVersions

Create a value of ListAssociationVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssociationVersions', listAssociationVersions_nextToken - A token to start the list. Use this token to get the next set of results.

$sel:maxResults:ListAssociationVersions', listAssociationVersions_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

$sel:associationId:ListAssociationVersions', listAssociationVersions_associationId - The association ID for which you want to view all versions.

Request Lenses

listAssociationVersions_nextToken :: Lens' ListAssociationVersions (Maybe Text)

A token to start the list. Use this token to get the next set of results.

listAssociationVersions_maxResults :: Lens' ListAssociationVersions (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

listAssociationVersions_associationId :: Lens' ListAssociationVersions Text

The association ID for which you want to view all versions.

Destructuring the Response

data ListAssociationVersionsResponse

See: newListAssociationVersionsResponse smart constructor.

Constructors

ListAssociationVersionsResponse'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. Use this token to get the next set of results.

associationVersions :: Maybe (NonEmpty AssociationVersionInfo)

Information about all versions of the association for the specified association ID.

httpStatus :: Int

The response's http status code.

newListAssociationVersionsResponse

Arguments

:: Int
$sel:httpStatus:ListAssociationVersionsResponse'

-> ListAssociationVersionsResponse

Create a value of ListAssociationVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssociationVersions', listAssociationVersionsResponse_nextToken - The token for the next set of items to return. Use this token to get the next set of results.

$sel:associationVersions:ListAssociationVersionsResponse', listAssociationVersionsResponse_associationVersions - Information about all versions of the association for the specified association ID.

$sel:httpStatus:ListAssociationVersionsResponse', listAssociationVersionsResponse_httpStatus - The response's http status code.

Response Lenses

listAssociationVersionsResponse_nextToken :: Lens' ListAssociationVersionsResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

listAssociationVersionsResponse_associationVersions :: Lens' ListAssociationVersionsResponse (Maybe (NonEmpty AssociationVersionInfo))

Information about all versions of the association for the specified association ID.

listAssociationVersionsResponse_httpStatus :: Lens' ListAssociationVersionsResponse Int

The response's http status code.