Network.AWS.SSM.ListAssociations

Description

Returns all State Manager associations in the current Amazon Web Services account and Amazon Web Services Region. You can limit the results to a specific State Manager association document or instance by specifying a filter. State Manager is a capability of Amazon Web Services Systems Manager.

This operation returns paginated results.

Creating a Request

data ListAssociations

See: newListAssociations smart constructor.

Constructors

ListAssociations'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

associationFilterList :: Maybe (NonEmpty AssociationFilter)

One or more filters. Use a filter to return a more specific list of results.

Filtering associations using the InstanceID attribute only returns legacy associations created using the InstanceID attribute. Associations targeting the instance that are part of the Target Attributes ResourceGroup or Tags aren't returned.

newListAssociations :: ListAssociations

Create a value of ListAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssociations', listAssociations_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:maxResults:ListAssociations', listAssociations_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

$sel:associationFilterList:ListAssociations', listAssociations_associationFilterList - One or more filters. Use a filter to return a more specific list of results.

Filtering associations using the InstanceID attribute only returns legacy associations created using the InstanceID attribute. Associations targeting the instance that are part of the Target Attributes ResourceGroup or Tags aren't returned.

Request Lenses

listAssociations_nextToken :: Lens' ListAssociations (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

listAssociations_maxResults :: Lens' ListAssociations (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

listAssociations_associationFilterList :: Lens' ListAssociations (Maybe (NonEmpty AssociationFilter))

One or more filters. Use a filter to return a more specific list of results.

Filtering associations using the InstanceID attribute only returns legacy associations created using the InstanceID attribute. Associations targeting the instance that are part of the Target Attributes ResourceGroup or Tags aren't returned.

Destructuring the Response

data ListAssociationsResponse

See: newListAssociationsResponse smart constructor.

Constructors

ListAssociationsResponse'

Fields

nextToken :: Maybe Text

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

associations :: Maybe [Association]

The associations.

httpStatus :: Int

The response's http status code.

newListAssociationsResponse

Arguments

:: Int
$sel:httpStatus:ListAssociationsResponse'

-> ListAssociationsResponse

Create a value of ListAssociationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssociations', listAssociationsResponse_nextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

$sel:associations:ListAssociationsResponse', listAssociationsResponse_associations - The associations.

$sel:httpStatus:ListAssociationsResponse', listAssociationsResponse_httpStatus - The response's http status code.

Response Lenses

listAssociationsResponse_nextToken :: Lens' ListAssociationsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

listAssociationsResponse_associations :: Lens' ListAssociationsResponse (Maybe [Association])

The associations.

listAssociationsResponse_httpStatus :: Lens' ListAssociationsResponse Int

The response's http status code.