Network.AWS.SSM.ListCommands

Description

Lists the commands requested by users of the Amazon Web Services account.

This operation returns paginated results.

Creating a Request

data ListCommands

See: newListCommands smart constructor.

Constructors

ListCommands'

Fields

instanceId :: Maybe Text

(Optional) Lists commands issued against this instance ID.

You can't specify an instance ID in the same command that you specify Status = Pending. This is because the command hasn't reached the instance yet.

filters :: Maybe (NonEmpty CommandFilter)

(Optional) One or more filters. Use a filter to return a more specific list of results.

nextToken :: Maybe Text

(Optional) The token for the next set of items to return. (You received this token from a previous call.)

commandId :: Maybe Text

(Optional) If provided, lists only the specified command.

maxResults :: Maybe Natural

(Optional) The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

newListCommands :: ListCommands

Create a value of ListCommands with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:ListCommands', listCommands_instanceId - (Optional) Lists commands issued against this instance ID.

You can't specify an instance ID in the same command that you specify Status = Pending. This is because the command hasn't reached the instance yet.

$sel:filters:ListCommands', listCommands_filters - (Optional) One or more filters. Use a filter to return a more specific list of results.

$sel:nextToken:ListCommands', listCommands_nextToken - (Optional) The token for the next set of items to return. (You received this token from a previous call.)

$sel:commandId:ListCommands', listCommands_commandId - (Optional) If provided, lists only the specified command.

$sel:maxResults:ListCommands', listCommands_maxResults - (Optional) The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

Request Lenses

listCommands_instanceId :: Lens' ListCommands (Maybe Text)

(Optional) Lists commands issued against this instance ID.

You can't specify an instance ID in the same command that you specify Status = Pending. This is because the command hasn't reached the instance yet.

listCommands_filters :: Lens' ListCommands (Maybe (NonEmpty CommandFilter))

(Optional) One or more filters. Use a filter to return a more specific list of results.

listCommands_nextToken :: Lens' ListCommands (Maybe Text)

(Optional) The token for the next set of items to return. (You received this token from a previous call.)

listCommands_commandId :: Lens' ListCommands (Maybe Text)

(Optional) If provided, lists only the specified command.

listCommands_maxResults :: Lens' ListCommands (Maybe Natural)

(Optional) The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

Destructuring the Response

data ListCommandsResponse

See: newListCommandsResponse smart constructor.

Constructors

ListCommandsResponse'

Fields

commands :: Maybe [Command]

(Optional) The list of commands requested by the user.

nextToken :: Maybe Text

(Optional) The token for the next set of items to return. (You received this token from a previous call.)

httpStatus :: Int

The response's http status code.

newListCommandsResponse

Arguments

:: Int
$sel:httpStatus:ListCommandsResponse'

-> ListCommandsResponse

Create a value of ListCommandsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:commands:ListCommandsResponse', listCommandsResponse_commands - (Optional) The list of commands requested by the user.

$sel:nextToken:ListCommands', listCommandsResponse_nextToken - (Optional) The token for the next set of items to return. (You received this token from a previous call.)

$sel:httpStatus:ListCommandsResponse', listCommandsResponse_httpStatus - The response's http status code.

Response Lenses

listCommandsResponse_commands :: Lens' ListCommandsResponse (Maybe [Command])

(Optional) The list of commands requested by the user.

listCommandsResponse_nextToken :: Lens' ListCommandsResponse (Maybe Text)

(Optional) The token for the next set of items to return. (You received this token from a previous call.)

listCommandsResponse_httpStatus :: Lens' ListCommandsResponse Int

The response's http status code.