Network.AWS.SSM.ListDocuments

Description

Returns all Systems Manager (SSM) documents in the current Amazon Web Services account and Amazon Web Services Region. You can limit the results of this request by using a filter.

This operation returns paginated results.

Creating a Request

data ListDocuments

See: newListDocuments smart constructor.

Constructors

ListDocuments'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

filters :: Maybe [DocumentKeyValuesFilter]

One or more DocumentKeyValuesFilter objects. Use a filter to return a more specific list of results. For keys, you can specify one or more key-value pair tags that have been applied to a document. Other valid keys include Owner, Name, PlatformTypes, DocumentType, and TargetType. For example, to return documents you own use Key=Owner,Values=Self. To specify a custom key-value pair, use the format Key=tag:tagName,Values=valueName.

This API operation only supports filtering documents by using a single tag key and one or more tag values. For example: Key=tag:tagName,Values=valueName1,valueName2

documentFilterList :: Maybe (NonEmpty DocumentFilter)

This data type is deprecated. Instead, use Filters.

newListDocuments :: ListDocuments

Create a value of ListDocuments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDocuments', listDocuments_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:maxResults:ListDocuments', listDocuments_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

$sel:filters:ListDocuments', listDocuments_filters - One or more DocumentKeyValuesFilter objects. Use a filter to return a more specific list of results. For keys, you can specify one or more key-value pair tags that have been applied to a document. Other valid keys include Owner, Name, PlatformTypes, DocumentType, and TargetType. For example, to return documents you own use Key=Owner,Values=Self. To specify a custom key-value pair, use the format Key=tag:tagName,Values=valueName.

This API operation only supports filtering documents by using a single tag key and one or more tag values. For example: Key=tag:tagName,Values=valueName1,valueName2

$sel:documentFilterList:ListDocuments', listDocuments_documentFilterList - This data type is deprecated. Instead, use Filters.

Request Lenses

listDocuments_nextToken :: Lens' ListDocuments (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

listDocuments_maxResults :: Lens' ListDocuments (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

listDocuments_filters :: Lens' ListDocuments (Maybe [DocumentKeyValuesFilter])

One or more DocumentKeyValuesFilter objects. Use a filter to return a more specific list of results. For keys, you can specify one or more key-value pair tags that have been applied to a document. Other valid keys include Owner, Name, PlatformTypes, DocumentType, and TargetType. For example, to return documents you own use Key=Owner,Values=Self. To specify a custom key-value pair, use the format Key=tag:tagName,Values=valueName.

This API operation only supports filtering documents by using a single tag key and one or more tag values. For example: Key=tag:tagName,Values=valueName1,valueName2

listDocuments_documentFilterList :: Lens' ListDocuments (Maybe (NonEmpty DocumentFilter))

This data type is deprecated. Instead, use Filters.

Destructuring the Response

data ListDocumentsResponse

See: newListDocumentsResponse smart constructor.

Constructors

ListDocumentsResponse'

Fields

nextToken :: Maybe Text

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

documentIdentifiers :: Maybe [DocumentIdentifier]

The names of the SSM documents.

httpStatus :: Int

The response's http status code.

newListDocumentsResponse

Arguments

:: Int
$sel:httpStatus:ListDocumentsResponse'

-> ListDocumentsResponse

Create a value of ListDocumentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDocuments', listDocumentsResponse_nextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

$sel:documentIdentifiers:ListDocumentsResponse', listDocumentsResponse_documentIdentifiers - The names of the SSM documents.

$sel:httpStatus:ListDocumentsResponse', listDocumentsResponse_httpStatus - The response's http status code.

Response Lenses

listDocumentsResponse_nextToken :: Lens' ListDocumentsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

listDocumentsResponse_documentIdentifiers :: Lens' ListDocumentsResponse (Maybe [DocumentIdentifier])

The names of the SSM documents.

listDocumentsResponse_httpStatus :: Lens' ListDocumentsResponse Int

The response's http status code.