Network.AWS.SSM.ModifyDocumentPermission

Description

Shares a Amazon Web Services Systems Manager document (SSM document)publicly or privately. If you share a document privately, you must specify the Amazon Web Services user account IDs for those people who can use the document. If you share a document publicly, you must specify All as the account ID.

Creating a Request

data ModifyDocumentPermission

See: newModifyDocumentPermission smart constructor.

Constructors

ModifyDocumentPermission'

Fields

accountIdsToAdd :: Maybe [Text]

The Amazon Web Services user accounts that should have access to the document. The account IDs can either be a group of account IDs or All.

sharedDocumentVersion :: Maybe Text

(Optional) The version of the document to share. If it isn't specified, the system choose the Default version to share.

accountIdsToRemove :: Maybe [Text]

The Amazon Web Services user accounts that should no longer have access to the document. The Amazon Web Services user account can either be a group of account IDs or All. This action has a higher priority than AccountIdsToAdd. If you specify an account ID to add and the same ID to remove, the system removes access to the document.

name :: Text

The name of the document that you want to share.

permissionType :: DocumentPermissionType

The permission type for the document. The permission type can be Share.

newModifyDocumentPermission

Arguments

:: Text
$sel:name:ModifyDocumentPermission'

-> DocumentPermissionType
$sel:permissionType:ModifyDocumentPermission'

-> ModifyDocumentPermission

Create a value of ModifyDocumentPermission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountIdsToAdd:ModifyDocumentPermission', modifyDocumentPermission_accountIdsToAdd - The Amazon Web Services user accounts that should have access to the document. The account IDs can either be a group of account IDs or All.

$sel:sharedDocumentVersion:ModifyDocumentPermission', modifyDocumentPermission_sharedDocumentVersion - (Optional) The version of the document to share. If it isn't specified, the system choose the Default version to share.

$sel:accountIdsToRemove:ModifyDocumentPermission', modifyDocumentPermission_accountIdsToRemove - The Amazon Web Services user accounts that should no longer have access to the document. The Amazon Web Services user account can either be a group of account IDs or All. This action has a higher priority than AccountIdsToAdd. If you specify an account ID to add and the same ID to remove, the system removes access to the document.

$sel:name:ModifyDocumentPermission', modifyDocumentPermission_name - The name of the document that you want to share.

$sel:permissionType:ModifyDocumentPermission', modifyDocumentPermission_permissionType - The permission type for the document. The permission type can be Share.

Request Lenses

modifyDocumentPermission_accountIdsToAdd :: Lens' ModifyDocumentPermission (Maybe [Text])

The Amazon Web Services user accounts that should have access to the document. The account IDs can either be a group of account IDs or All.

modifyDocumentPermission_sharedDocumentVersion :: Lens' ModifyDocumentPermission (Maybe Text)

(Optional) The version of the document to share. If it isn't specified, the system choose the Default version to share.

modifyDocumentPermission_accountIdsToRemove :: Lens' ModifyDocumentPermission (Maybe [Text])

The Amazon Web Services user accounts that should no longer have access to the document. The Amazon Web Services user account can either be a group of account IDs or All. This action has a higher priority than AccountIdsToAdd. If you specify an account ID to add and the same ID to remove, the system removes access to the document.

modifyDocumentPermission_name :: Lens' ModifyDocumentPermission Text

The name of the document that you want to share.

modifyDocumentPermission_permissionType :: Lens' ModifyDocumentPermission DocumentPermissionType

The permission type for the document. The permission type can be Share.

Destructuring the Response

data ModifyDocumentPermissionResponse

See: newModifyDocumentPermissionResponse smart constructor.

Constructors

ModifyDocumentPermissionResponse'

Fields

httpStatus :: Int

The response's http status code.

newModifyDocumentPermissionResponse

Arguments

:: Int
$sel:httpStatus:ModifyDocumentPermissionResponse'

-> ModifyDocumentPermissionResponse

Create a value of ModifyDocumentPermissionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ModifyDocumentPermissionResponse', modifyDocumentPermissionResponse_httpStatus - The response's http status code.

Response Lenses

modifyDocumentPermissionResponse_httpStatus :: Lens' ModifyDocumentPermissionResponse Int

The response's http status code.