Network.AWS.SSM.RegisterDefaultPatchBaseline

Description

Defines the default patch baseline for the relevant operating system.

To reset the Amazon Web Services-predefined patch baseline as the default, specify the full patch baseline Amazon Resource Name (ARN) as the baseline ID value. For example, for CentOS, specify arn:aws:ssm:us-east-2:733109147000:patchbaseline/pb-0574b43a65ea646ed instead of pb-0574b43a65ea646ed.

Creating a Request

data RegisterDefaultPatchBaseline

See: newRegisterDefaultPatchBaseline smart constructor.

Constructors

RegisterDefaultPatchBaseline'

Fields

baselineId :: Text

The ID of the patch baseline that should be the default patch baseline.

newRegisterDefaultPatchBaseline

Arguments

:: Text
$sel:baselineId:RegisterDefaultPatchBaseline'

-> RegisterDefaultPatchBaseline

Create a value of RegisterDefaultPatchBaseline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baselineId:RegisterDefaultPatchBaseline', registerDefaultPatchBaseline_baselineId - The ID of the patch baseline that should be the default patch baseline.

Request Lenses

registerDefaultPatchBaseline_baselineId :: Lens' RegisterDefaultPatchBaseline Text

The ID of the patch baseline that should be the default patch baseline.

Destructuring the Response

data RegisterDefaultPatchBaselineResponse

See: newRegisterDefaultPatchBaselineResponse smart constructor.

Constructors

RegisterDefaultPatchBaselineResponse'

Fields

baselineId :: Maybe Text

The ID of the default patch baseline.

httpStatus :: Int

The response's http status code.

newRegisterDefaultPatchBaselineResponse

Arguments

:: Int
$sel:httpStatus:RegisterDefaultPatchBaselineResponse'

-> RegisterDefaultPatchBaselineResponse

Create a value of RegisterDefaultPatchBaselineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baselineId:RegisterDefaultPatchBaseline', registerDefaultPatchBaselineResponse_baselineId - The ID of the default patch baseline.

$sel:httpStatus:RegisterDefaultPatchBaselineResponse', registerDefaultPatchBaselineResponse_httpStatus - The response's http status code.

Response Lenses

registerDefaultPatchBaselineResponse_baselineId :: Lens' RegisterDefaultPatchBaselineResponse (Maybe Text)

The ID of the default patch baseline.

registerDefaultPatchBaselineResponse_httpStatus :: Lens' RegisterDefaultPatchBaselineResponse Int

The response's http status code.