Network.AWS.SSM.SendAutomationSignal

Description

Sends a signal to an Automation execution to change the current behavior or status of the execution.

Creating a Request

data SendAutomationSignal

See: newSendAutomationSignal smart constructor.

Constructors

SendAutomationSignal'

Fields

payload :: Maybe (HashMap Text [Text])

The data sent with the signal. The data schema depends on the type of signal used in the request.

For Approve and Reject signal types, the payload is an optional comment that you can send with the signal type. For example:

Comment="Looks good"

For StartStep and Resume signal types, you must send the name of the Automation step to start or resume as the payload. For example:

StepName="step1"

For the StopStep signal type, you must send the step execution ID as the payload. For example:

StepExecutionId="97fff367-fc5a-4299-aed8-0123456789ab"

automationExecutionId :: Text

The unique identifier for an existing Automation execution that you want to send the signal to.

signalType :: SignalType

The type of signal to send to an Automation execution.

newSendAutomationSignal

Arguments

:: Text
$sel:automationExecutionId:SendAutomationSignal'

-> SignalType
$sel:signalType:SendAutomationSignal'

-> SendAutomationSignal

Create a value of SendAutomationSignal with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:payload:SendAutomationSignal', sendAutomationSignal_payload - The data sent with the signal. The data schema depends on the type of signal used in the request.

For Approve and Reject signal types, the payload is an optional comment that you can send with the signal type. For example:

Comment="Looks good"

For StartStep and Resume signal types, you must send the name of the Automation step to start or resume as the payload. For example:

StepName="step1"

For the StopStep signal type, you must send the step execution ID as the payload. For example:

StepExecutionId="97fff367-fc5a-4299-aed8-0123456789ab"

$sel:automationExecutionId:SendAutomationSignal', sendAutomationSignal_automationExecutionId - The unique identifier for an existing Automation execution that you want to send the signal to.

$sel:signalType:SendAutomationSignal', sendAutomationSignal_signalType - The type of signal to send to an Automation execution.

Request Lenses

sendAutomationSignal_payload :: Lens' SendAutomationSignal (Maybe (HashMap Text [Text]))

The data sent with the signal. The data schema depends on the type of signal used in the request.

For Approve and Reject signal types, the payload is an optional comment that you can send with the signal type. For example:

Comment="Looks good"

For StartStep and Resume signal types, you must send the name of the Automation step to start or resume as the payload. For example:

StepName="step1"

For the StopStep signal type, you must send the step execution ID as the payload. For example:

StepExecutionId="97fff367-fc5a-4299-aed8-0123456789ab"

sendAutomationSignal_automationExecutionId :: Lens' SendAutomationSignal Text

The unique identifier for an existing Automation execution that you want to send the signal to.

sendAutomationSignal_signalType :: Lens' SendAutomationSignal SignalType

The type of signal to send to an Automation execution.

Destructuring the Response

data SendAutomationSignalResponse

See: newSendAutomationSignalResponse smart constructor.

Constructors

SendAutomationSignalResponse'

Fields

httpStatus :: Int

The response's http status code.

newSendAutomationSignalResponse

Arguments

:: Int
$sel:httpStatus:SendAutomationSignalResponse'

-> SendAutomationSignalResponse

Create a value of SendAutomationSignalResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:SendAutomationSignalResponse', sendAutomationSignalResponse_httpStatus - The response's http status code.

Response Lenses

sendAutomationSignalResponse_httpStatus :: Lens' SendAutomationSignalResponse Int

The response's http status code.