Network.AWS.SSM.SendCommand

Description

Runs commands on one or more managed instances.

Creating a Request

data SendCommand

See: newSendCommand smart constructor.

Constructors

SendCommand'

Fields

instanceIds :: Maybe [Text]

The IDs of the instances where the command should run. Specifying instance IDs is most useful when you are targeting a limited number of instances, though you can specify up to 50 IDs.

To target a larger number of instances, or if you prefer not to list individual instance IDs, we recommend using the Targets option instead. Using Targets, which accepts tag key-value pairs to identify the instances to send commands to, you can a send command to tens, hundreds, or thousands of instances at once.

For more information about how to use targets, see Using targets and rate controls to send commands to a fleet in the Amazon Web Services Systems Manager User Guide.

maxErrors :: Maybe Text

The maximum number of errors allowed without the command failing. When the command fails one more time beyond the value of MaxErrors, the systems stops sending the command to additional targets. You can specify a number like 10 or a percentage like 10%. The default value is 0. For more information about how to use MaxErrors, see Using error controls in the Amazon Web Services Systems Manager User Guide.

notificationConfig :: Maybe NotificationConfig

Configurations for sending notifications.

serviceRoleArn :: Maybe Text

The ARN of the Identity and Access Management (IAM) service role to use to publish Amazon Simple Notification Service (Amazon SNS) notifications for Run Command commands.

outputS3BucketName :: Maybe Text

The name of the S3 bucket where command execution responses should be stored.

comment :: Maybe Text

User-specified information about the command, such as a brief description of what the command should do.

documentHash :: Maybe Text

The Sha256 or Sha1 hash created by the system when the document was created.

Sha1 hashes have been deprecated.

targets :: Maybe [Target]

An array of search criteria that targets instances using a Key,Value combination that you specify. Specifying targets is most useful when you want to send a command to a large number of instances at once. Using Targets, which accepts tag key-value pairs to identify instances, you can send a command to tens, hundreds, or thousands of instances at once.

To send a command to a smaller number of instances, you can use the InstanceIds option instead.

For more information about how to use targets, see Sending commands to a fleet in the Amazon Web Services Systems Manager User Guide.

outputS3Region :: Maybe Text

(Deprecated) You can no longer specify this parameter. The system ignores it. Instead, Systems Manager automatically determines the Amazon Web Services Region of the S3 bucket.

maxConcurrency :: Maybe Text

(Optional) The maximum number of instances that are allowed to run the command at the same time. You can specify a number such as 10 or a percentage such as 10%. The default value is 50. For more information about how to use MaxConcurrency, see Using concurrency controls in the Amazon Web Services Systems Manager User Guide.

timeoutSeconds :: Maybe Natural

If this time is reached and the command hasn't already started running, it won't run.

outputS3KeyPrefix :: Maybe Text

The directory structure within the S3 bucket where the responses should be stored.

cloudWatchOutputConfig :: Maybe CloudWatchOutputConfig

Enables Amazon Web Services Systems Manager to send Run Command output to Amazon CloudWatch Logs. Run Command is a capability of Amazon Web Services Systems Manager.

documentVersion :: Maybe Text

The SSM document version to use in the request. You can specify $DEFAULT, $LATEST, or a specific version number. If you run commands by using the Command Line Interface (Amazon Web Services CLI), then you must escape the first two options by using a backslash. If you specify a version number, then you don't need to use the backslash. For example:

-document-version "\$DEFAULT"

-document-version "\$LATEST"

-document-version "3"

documentHashType :: Maybe DocumentHashType

Sha256 or Sha1.

Sha1 hashes have been deprecated.

parameters :: Maybe (HashMap Text [Text])

The required and optional parameters specified in the document being run.

documentName :: Text

The name of the Amazon Web Services Systems Manager document (SSM document) to run. This can be a public document or a custom document. To run a shared document belonging to another account, specify the document Amazon Resource Name (ARN). For more information about how to use shared documents, see Using shared SSM documents in the Amazon Web Services Systems Manager User Guide.

If you specify a document name or ARN that hasn't been shared with your account, you receive an InvalidDocument error.

newSendCommand

Arguments

:: Text
$sel:documentName:SendCommand'

-> SendCommand

Create a value of SendCommand with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceIds:SendCommand', sendCommand_instanceIds - The IDs of the instances where the command should run. Specifying instance IDs is most useful when you are targeting a limited number of instances, though you can specify up to 50 IDs.

To target a larger number of instances, or if you prefer not to list individual instance IDs, we recommend using the Targets option instead. Using Targets, which accepts tag key-value pairs to identify the instances to send commands to, you can a send command to tens, hundreds, or thousands of instances at once.

For more information about how to use targets, see Using targets and rate controls to send commands to a fleet in the Amazon Web Services Systems Manager User Guide.

$sel:maxErrors:SendCommand', sendCommand_maxErrors - The maximum number of errors allowed without the command failing. When the command fails one more time beyond the value of MaxErrors, the systems stops sending the command to additional targets. You can specify a number like 10 or a percentage like 10%. The default value is 0. For more information about how to use MaxErrors, see Using error controls in the Amazon Web Services Systems Manager User Guide.

$sel:notificationConfig:SendCommand', sendCommand_notificationConfig - Configurations for sending notifications.

$sel:serviceRoleArn:SendCommand', sendCommand_serviceRoleArn - The ARN of the Identity and Access Management (IAM) service role to use to publish Amazon Simple Notification Service (Amazon SNS) notifications for Run Command commands.

$sel:outputS3BucketName:SendCommand', sendCommand_outputS3BucketName - The name of the S3 bucket where command execution responses should be stored.

$sel:comment:SendCommand', sendCommand_comment - User-specified information about the command, such as a brief description of what the command should do.

$sel:documentHash:SendCommand', sendCommand_documentHash - The Sha256 or Sha1 hash created by the system when the document was created.

Sha1 hashes have been deprecated.

$sel:targets:SendCommand', sendCommand_targets - An array of search criteria that targets instances using a Key,Value combination that you specify. Specifying targets is most useful when you want to send a command to a large number of instances at once. Using Targets, which accepts tag key-value pairs to identify instances, you can send a command to tens, hundreds, or thousands of instances at once.

To send a command to a smaller number of instances, you can use the InstanceIds option instead.

For more information about how to use targets, see Sending commands to a fleet in the Amazon Web Services Systems Manager User Guide.

$sel:outputS3Region:SendCommand', sendCommand_outputS3Region - (Deprecated) You can no longer specify this parameter. The system ignores it. Instead, Systems Manager automatically determines the Amazon Web Services Region of the S3 bucket.

$sel:maxConcurrency:SendCommand', sendCommand_maxConcurrency - (Optional) The maximum number of instances that are allowed to run the command at the same time. You can specify a number such as 10 or a percentage such as 10%. The default value is 50. For more information about how to use MaxConcurrency, see Using concurrency controls in the Amazon Web Services Systems Manager User Guide.

$sel:timeoutSeconds:SendCommand', sendCommand_timeoutSeconds - If this time is reached and the command hasn't already started running, it won't run.

$sel:outputS3KeyPrefix:SendCommand', sendCommand_outputS3KeyPrefix - The directory structure within the S3 bucket where the responses should be stored.

$sel:cloudWatchOutputConfig:SendCommand', sendCommand_cloudWatchOutputConfig - Enables Amazon Web Services Systems Manager to send Run Command output to Amazon CloudWatch Logs. Run Command is a capability of Amazon Web Services Systems Manager.

$sel:documentVersion:SendCommand', sendCommand_documentVersion - The SSM document version to use in the request. You can specify $DEFAULT, $LATEST, or a specific version number. If you run commands by using the Command Line Interface (Amazon Web Services CLI), then you must escape the first two options by using a backslash. If you specify a version number, then you don't need to use the backslash. For example:

-document-version "\$DEFAULT"

-document-version "\$LATEST"

-document-version "3"

$sel:documentHashType:SendCommand', sendCommand_documentHashType - Sha256 or Sha1.

Sha1 hashes have been deprecated.

$sel:parameters:SendCommand', sendCommand_parameters - The required and optional parameters specified in the document being run.

$sel:documentName:SendCommand', sendCommand_documentName - The name of the Amazon Web Services Systems Manager document (SSM document) to run. This can be a public document or a custom document. To run a shared document belonging to another account, specify the document Amazon Resource Name (ARN). For more information about how to use shared documents, see Using shared SSM documents in the Amazon Web Services Systems Manager User Guide.

If you specify a document name or ARN that hasn't been shared with your account, you receive an InvalidDocument error.

Request Lenses

sendCommand_instanceIds :: Lens' SendCommand (Maybe [Text])

The IDs of the instances where the command should run. Specifying instance IDs is most useful when you are targeting a limited number of instances, though you can specify up to 50 IDs.

To target a larger number of instances, or if you prefer not to list individual instance IDs, we recommend using the Targets option instead. Using Targets, which accepts tag key-value pairs to identify the instances to send commands to, you can a send command to tens, hundreds, or thousands of instances at once.

For more information about how to use targets, see Using targets and rate controls to send commands to a fleet in the Amazon Web Services Systems Manager User Guide.

sendCommand_maxErrors :: Lens' SendCommand (Maybe Text)

The maximum number of errors allowed without the command failing. When the command fails one more time beyond the value of MaxErrors, the systems stops sending the command to additional targets. You can specify a number like 10 or a percentage like 10%. The default value is 0. For more information about how to use MaxErrors, see Using error controls in the Amazon Web Services Systems Manager User Guide.

sendCommand_notificationConfig :: Lens' SendCommand (Maybe NotificationConfig)

Configurations for sending notifications.

sendCommand_serviceRoleArn :: Lens' SendCommand (Maybe Text)

The ARN of the Identity and Access Management (IAM) service role to use to publish Amazon Simple Notification Service (Amazon SNS) notifications for Run Command commands.

sendCommand_outputS3BucketName :: Lens' SendCommand (Maybe Text)

The name of the S3 bucket where command execution responses should be stored.

sendCommand_comment :: Lens' SendCommand (Maybe Text)

User-specified information about the command, such as a brief description of what the command should do.

sendCommand_documentHash :: Lens' SendCommand (Maybe Text)

The Sha256 or Sha1 hash created by the system when the document was created.

Sha1 hashes have been deprecated.

sendCommand_targets :: Lens' SendCommand (Maybe [Target])

An array of search criteria that targets instances using a Key,Value combination that you specify. Specifying targets is most useful when you want to send a command to a large number of instances at once. Using Targets, which accepts tag key-value pairs to identify instances, you can send a command to tens, hundreds, or thousands of instances at once.

To send a command to a smaller number of instances, you can use the InstanceIds option instead.

For more information about how to use targets, see Sending commands to a fleet in the Amazon Web Services Systems Manager User Guide.

sendCommand_outputS3Region :: Lens' SendCommand (Maybe Text)

(Deprecated) You can no longer specify this parameter. The system ignores it. Instead, Systems Manager automatically determines the Amazon Web Services Region of the S3 bucket.

sendCommand_maxConcurrency :: Lens' SendCommand (Maybe Text)

(Optional) The maximum number of instances that are allowed to run the command at the same time. You can specify a number such as 10 or a percentage such as 10%. The default value is 50. For more information about how to use MaxConcurrency, see Using concurrency controls in the Amazon Web Services Systems Manager User Guide.

sendCommand_timeoutSeconds :: Lens' SendCommand (Maybe Natural)

If this time is reached and the command hasn't already started running, it won't run.

sendCommand_outputS3KeyPrefix :: Lens' SendCommand (Maybe Text)

The directory structure within the S3 bucket where the responses should be stored.

sendCommand_cloudWatchOutputConfig :: Lens' SendCommand (Maybe CloudWatchOutputConfig)

Enables Amazon Web Services Systems Manager to send Run Command output to Amazon CloudWatch Logs. Run Command is a capability of Amazon Web Services Systems Manager.

sendCommand_documentVersion :: Lens' SendCommand (Maybe Text)

The SSM document version to use in the request. You can specify $DEFAULT, $LATEST, or a specific version number. If you run commands by using the Command Line Interface (Amazon Web Services CLI), then you must escape the first two options by using a backslash. If you specify a version number, then you don't need to use the backslash. For example:

-document-version "\$DEFAULT"

-document-version "\$LATEST"

-document-version "3"

sendCommand_documentHashType :: Lens' SendCommand (Maybe DocumentHashType)

Sha256 or Sha1.

Sha1 hashes have been deprecated.

sendCommand_parameters :: Lens' SendCommand (Maybe (HashMap Text [Text]))

The required and optional parameters specified in the document being run.

sendCommand_documentName :: Lens' SendCommand Text

The name of the Amazon Web Services Systems Manager document (SSM document) to run. This can be a public document or a custom document. To run a shared document belonging to another account, specify the document Amazon Resource Name (ARN). For more information about how to use shared documents, see Using shared SSM documents in the Amazon Web Services Systems Manager User Guide.

If you specify a document name or ARN that hasn't been shared with your account, you receive an InvalidDocument error.

Destructuring the Response

data SendCommandResponse

See: newSendCommandResponse smart constructor.

Constructors

SendCommandResponse'

Fields

command :: Maybe Command

The request as it was received by Systems Manager. Also provides the command ID which can be used future references to this request.

httpStatus :: Int

The response's http status code.

newSendCommandResponse

Arguments

:: Int
$sel:httpStatus:SendCommandResponse'

-> SendCommandResponse

Create a value of SendCommandResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:command:SendCommandResponse', sendCommandResponse_command - The request as it was received by Systems Manager. Also provides the command ID which can be used future references to this request.

$sel:httpStatus:SendCommandResponse', sendCommandResponse_httpStatus - The response's http status code.

Response Lenses

sendCommandResponse_command :: Lens' SendCommandResponse (Maybe Command)

The request as it was received by Systems Manager. Also provides the command ID which can be used future references to this request.

sendCommandResponse_httpStatus :: Lens' SendCommandResponse Int

The response's http status code.