Network.AWS.SSM.StartAutomationExecution

Description

Initiates execution of an Automation runbook.

Creating a Request

data StartAutomationExecution

See: newStartAutomationExecution smart constructor.

Constructors

StartAutomationExecution'

Fields

maxErrors :: Maybe Text

The number of errors that are allowed before the system stops running the automation on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops running the automation when the fourth error is received. If you specify 0, then the system stops running the automation on additional targets after the first error result is returned. If you run an automation on 50 resources and set max-errors to 10%, then the system stops running the automation on additional targets when the sixth error is received.

Executions that are already running an automation when max-errors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set max-concurrency to 1 so the executions proceed one at a time.

mode :: Maybe ExecutionMode

The execution mode of the automation. Valid modes include the following: Auto and Interactive. The default mode is Auto.

targets :: Maybe [Target]

A key-value mapping to target resources. Required if you specify TargetParameterName.

targetLocations :: Maybe (NonEmpty TargetLocation)

A location is a combination of Amazon Web Services Regions and/or Amazon Web Services accounts where you want to run the automation. Use this operation to start an automation in multiple Amazon Web Services Regions and multiple Amazon Web Services accounts. For more information, see Running Automation workflows in multiple Amazon Web Services Regions and Amazon Web Services accounts in the Amazon Web Services Systems Manager User Guide.

targetParameterName :: Maybe Text

The name of the parameter used as the target resource for the rate-controlled execution. Required if you specify targets.

tags :: Maybe [Tag]

Optional metadata that you assign to a resource. You can specify a maximum of five tags for an automation. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag an automation to identify an environment or operating system. In this case, you could specify the following key-value pairs:

Key=environment,Value=test

Key=OS,Value=Windows

To add tags to an existing patch baseline, use the AddTagsToResource operation.

maxConcurrency :: Maybe Text

The maximum number of targets allowed to run this task in parallel. You can specify a number, such as 10, or a percentage, such as 10%. The default value is 10.

targetMaps :: Maybe [HashMap Text [Text]]

A key-value mapping of document parameters to target resources. Both Targets and TargetMaps can't be specified together.

documentVersion :: Maybe Text

The version of the Automation runbook to use for this execution.

clientToken :: Maybe Text

User-provided idempotency token. The token must be unique, is case insensitive, enforces the UUID format, and can't be reused.

parameters :: Maybe (HashMap Text [Text])

A key-value map of execution parameters, which match the declared parameters in the Automation runbook.

documentName :: Text

The name of the SSM document to run. This can be a public document or a custom document. To run a shared document belonging to another account, specify the document ARN. For more information about how to use shared documents, see Using shared SSM documents in the Amazon Web Services Systems Manager User Guide.

newStartAutomationExecution

Arguments

:: Text
$sel:documentName:StartAutomationExecution'

-> StartAutomationExecution

Create a value of StartAutomationExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxErrors:StartAutomationExecution', startAutomationExecution_maxErrors - The number of errors that are allowed before the system stops running the automation on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops running the automation when the fourth error is received. If you specify 0, then the system stops running the automation on additional targets after the first error result is returned. If you run an automation on 50 resources and set max-errors to 10%, then the system stops running the automation on additional targets when the sixth error is received.

Executions that are already running an automation when max-errors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set max-concurrency to 1 so the executions proceed one at a time.

$sel:mode:StartAutomationExecution', startAutomationExecution_mode - The execution mode of the automation. Valid modes include the following: Auto and Interactive. The default mode is Auto.

$sel:targets:StartAutomationExecution', startAutomationExecution_targets - A key-value mapping to target resources. Required if you specify TargetParameterName.

$sel:targetLocations:StartAutomationExecution', startAutomationExecution_targetLocations - A location is a combination of Amazon Web Services Regions and/or Amazon Web Services accounts where you want to run the automation. Use this operation to start an automation in multiple Amazon Web Services Regions and multiple Amazon Web Services accounts. For more information, see Running Automation workflows in multiple Amazon Web Services Regions and Amazon Web Services accounts in the Amazon Web Services Systems Manager User Guide.

$sel:targetParameterName:StartAutomationExecution', startAutomationExecution_targetParameterName - The name of the parameter used as the target resource for the rate-controlled execution. Required if you specify targets.

$sel:tags:StartAutomationExecution', startAutomationExecution_tags - Optional metadata that you assign to a resource. You can specify a maximum of five tags for an automation. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag an automation to identify an environment or operating system. In this case, you could specify the following key-value pairs:

Key=environment,Value=test

Key=OS,Value=Windows

To add tags to an existing patch baseline, use the AddTagsToResource operation.

$sel:maxConcurrency:StartAutomationExecution', startAutomationExecution_maxConcurrency - The maximum number of targets allowed to run this task in parallel. You can specify a number, such as 10, or a percentage, such as 10%. The default value is 10.

$sel:targetMaps:StartAutomationExecution', startAutomationExecution_targetMaps - A key-value mapping of document parameters to target resources. Both Targets and TargetMaps can't be specified together.

$sel:documentVersion:StartAutomationExecution', startAutomationExecution_documentVersion - The version of the Automation runbook to use for this execution.

$sel:clientToken:StartAutomationExecution', startAutomationExecution_clientToken - User-provided idempotency token. The token must be unique, is case insensitive, enforces the UUID format, and can't be reused.

$sel:parameters:StartAutomationExecution', startAutomationExecution_parameters - A key-value map of execution parameters, which match the declared parameters in the Automation runbook.

$sel:documentName:StartAutomationExecution', startAutomationExecution_documentName - The name of the SSM document to run. This can be a public document or a custom document. To run a shared document belonging to another account, specify the document ARN. For more information about how to use shared documents, see Using shared SSM documents in the Amazon Web Services Systems Manager User Guide.

Request Lenses

startAutomationExecution_maxErrors :: Lens' StartAutomationExecution (Maybe Text)

The number of errors that are allowed before the system stops running the automation on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops running the automation when the fourth error is received. If you specify 0, then the system stops running the automation on additional targets after the first error result is returned. If you run an automation on 50 resources and set max-errors to 10%, then the system stops running the automation on additional targets when the sixth error is received.

Executions that are already running an automation when max-errors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set max-concurrency to 1 so the executions proceed one at a time.

startAutomationExecution_mode :: Lens' StartAutomationExecution (Maybe ExecutionMode)

The execution mode of the automation. Valid modes include the following: Auto and Interactive. The default mode is Auto.

startAutomationExecution_targets :: Lens' StartAutomationExecution (Maybe [Target])

A key-value mapping to target resources. Required if you specify TargetParameterName.

startAutomationExecution_targetLocations :: Lens' StartAutomationExecution (Maybe (NonEmpty TargetLocation))

A location is a combination of Amazon Web Services Regions and/or Amazon Web Services accounts where you want to run the automation. Use this operation to start an automation in multiple Amazon Web Services Regions and multiple Amazon Web Services accounts. For more information, see Running Automation workflows in multiple Amazon Web Services Regions and Amazon Web Services accounts in the Amazon Web Services Systems Manager User Guide.

startAutomationExecution_targetParameterName :: Lens' StartAutomationExecution (Maybe Text)

The name of the parameter used as the target resource for the rate-controlled execution. Required if you specify targets.

startAutomationExecution_tags :: Lens' StartAutomationExecution (Maybe [Tag])

Optional metadata that you assign to a resource. You can specify a maximum of five tags for an automation. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag an automation to identify an environment or operating system. In this case, you could specify the following key-value pairs:

Key=environment,Value=test

Key=OS,Value=Windows

To add tags to an existing patch baseline, use the AddTagsToResource operation.

startAutomationExecution_maxConcurrency :: Lens' StartAutomationExecution (Maybe Text)

The maximum number of targets allowed to run this task in parallel. You can specify a number, such as 10, or a percentage, such as 10%. The default value is 10.

startAutomationExecution_targetMaps :: Lens' StartAutomationExecution (Maybe [HashMap Text [Text]])

A key-value mapping of document parameters to target resources. Both Targets and TargetMaps can't be specified together.

startAutomationExecution_documentVersion :: Lens' StartAutomationExecution (Maybe Text)

The version of the Automation runbook to use for this execution.

startAutomationExecution_clientToken :: Lens' StartAutomationExecution (Maybe Text)

User-provided idempotency token. The token must be unique, is case insensitive, enforces the UUID format, and can't be reused.

startAutomationExecution_parameters :: Lens' StartAutomationExecution (Maybe (HashMap Text [Text]))

A key-value map of execution parameters, which match the declared parameters in the Automation runbook.

startAutomationExecution_documentName :: Lens' StartAutomationExecution Text

The name of the SSM document to run. This can be a public document or a custom document. To run a shared document belonging to another account, specify the document ARN. For more information about how to use shared documents, see Using shared SSM documents in the Amazon Web Services Systems Manager User Guide.

Destructuring the Response

data StartAutomationExecutionResponse

See: newStartAutomationExecutionResponse smart constructor.

Constructors

StartAutomationExecutionResponse'

Fields

automationExecutionId :: Maybe Text

The unique ID of a newly scheduled automation execution.

httpStatus :: Int

The response's http status code.

newStartAutomationExecutionResponse

Arguments

:: Int
$sel:httpStatus:StartAutomationExecutionResponse'

-> StartAutomationExecutionResponse

Create a value of StartAutomationExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:automationExecutionId:StartAutomationExecutionResponse', startAutomationExecutionResponse_automationExecutionId - The unique ID of a newly scheduled automation execution.

$sel:httpStatus:StartAutomationExecutionResponse', startAutomationExecutionResponse_httpStatus - The response's http status code.

Response Lenses

startAutomationExecutionResponse_automationExecutionId :: Lens' StartAutomationExecutionResponse (Maybe Text)

The unique ID of a newly scheduled automation execution.

startAutomationExecutionResponse_httpStatus :: Lens' StartAutomationExecutionResponse Int

The response's http status code.