Network.AWS.SSM.StartChangeRequestExecution

Description

Creates a change request for Change Manager. The Automation runbooks specified in the change request run only after all required approvals for the change request have been received.

Creating a Request

data StartChangeRequestExecution

See: newStartChangeRequestExecution smart constructor.

Constructors

StartChangeRequestExecution'

Fields

scheduledTime :: Maybe POSIX

The date and time specified in the change request to run the Automation runbooks.

The Automation runbooks specified for the runbook workflow can't run until all required approvals for the change request have been received.

changeDetails :: Maybe Text

User-provided details about the change. If no details are provided, content specified in the Template information section of the associated change template is added.

clientToken :: Maybe Text

The user-provided idempotency token. The token must be unique, is case insensitive, enforces the UUID format, and can't be reused.

autoApprove :: Maybe Bool

Indicates whether the change request can be approved automatically without the need for manual approvals.

If AutoApprovable is enabled in a change template, then setting AutoApprove to true in StartChangeRequestExecution creates a change request that bypasses approver review.

Change Calendar restrictions are not bypassed in this scenario. If the state of an associated calendar is CLOSED, change freeze approvers must still grant permission for this change request to run. If they don't, the change won't be processed until the calendar state is again OPEN.

scheduledEndTime :: Maybe POSIX

The time that the requester expects the runbook workflow related to the change request to complete. The time is an estimate only that the requester provides for reviewers.

parameters :: Maybe (HashMap Text [Text])

A key-value map of parameters that match the declared parameters in the change template document.

documentVersion :: Maybe Text

The version of the change template document to run during the runbook workflow.

changeRequestName :: Maybe Text

The name of the change request associated with the runbook workflow to be run.

tags :: Maybe [Tag]

Optional metadata that you assign to a resource. You can specify a maximum of five tags for a change request. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag a change request to identify an environment or target Amazon Web Services Region. In this case, you could specify the following key-value pairs:

Key=Environment,Value=Production

Key=Region,Value=us-east-2

documentName :: Text

The name of the change template document to run during the runbook workflow.

runbooks :: NonEmpty Runbook

Information about the Automation runbooks that are run during the runbook workflow.

The Automation runbooks specified for the runbook workflow can't run until all required approvals for the change request have been received.

newStartChangeRequestExecution

Arguments

:: Text
$sel:documentName:StartChangeRequestExecution'

-> NonEmpty Runbook
$sel:runbooks:StartChangeRequestExecution'

-> StartChangeRequestExecution

Create a value of StartChangeRequestExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scheduledTime:StartChangeRequestExecution', startChangeRequestExecution_scheduledTime - The date and time specified in the change request to run the Automation runbooks.

The Automation runbooks specified for the runbook workflow can't run until all required approvals for the change request have been received.

$sel:changeDetails:StartChangeRequestExecution', startChangeRequestExecution_changeDetails - User-provided details about the change. If no details are provided, content specified in the Template information section of the associated change template is added.

$sel:clientToken:StartChangeRequestExecution', startChangeRequestExecution_clientToken - The user-provided idempotency token. The token must be unique, is case insensitive, enforces the UUID format, and can't be reused.

$sel:autoApprove:StartChangeRequestExecution', startChangeRequestExecution_autoApprove - Indicates whether the change request can be approved automatically without the need for manual approvals.

If AutoApprovable is enabled in a change template, then setting AutoApprove to true in StartChangeRequestExecution creates a change request that bypasses approver review.

Change Calendar restrictions are not bypassed in this scenario. If the state of an associated calendar is CLOSED, change freeze approvers must still grant permission for this change request to run. If they don't, the change won't be processed until the calendar state is again OPEN.

$sel:scheduledEndTime:StartChangeRequestExecution', startChangeRequestExecution_scheduledEndTime - The time that the requester expects the runbook workflow related to the change request to complete. The time is an estimate only that the requester provides for reviewers.

$sel:parameters:StartChangeRequestExecution', startChangeRequestExecution_parameters - A key-value map of parameters that match the declared parameters in the change template document.

$sel:documentVersion:StartChangeRequestExecution', startChangeRequestExecution_documentVersion - The version of the change template document to run during the runbook workflow.

$sel:changeRequestName:StartChangeRequestExecution', startChangeRequestExecution_changeRequestName - The name of the change request associated with the runbook workflow to be run.

$sel:tags:StartChangeRequestExecution', startChangeRequestExecution_tags - Optional metadata that you assign to a resource. You can specify a maximum of five tags for a change request. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag a change request to identify an environment or target Amazon Web Services Region. In this case, you could specify the following key-value pairs:

Key=Environment,Value=Production

Key=Region,Value=us-east-2

$sel:documentName:StartChangeRequestExecution', startChangeRequestExecution_documentName - The name of the change template document to run during the runbook workflow.

$sel:runbooks:StartChangeRequestExecution', startChangeRequestExecution_runbooks - Information about the Automation runbooks that are run during the runbook workflow.

The Automation runbooks specified for the runbook workflow can't run until all required approvals for the change request have been received.

Request Lenses

startChangeRequestExecution_scheduledTime :: Lens' StartChangeRequestExecution (Maybe UTCTime)

The date and time specified in the change request to run the Automation runbooks.

The Automation runbooks specified for the runbook workflow can't run until all required approvals for the change request have been received.

startChangeRequestExecution_changeDetails :: Lens' StartChangeRequestExecution (Maybe Text)

User-provided details about the change. If no details are provided, content specified in the Template information section of the associated change template is added.

startChangeRequestExecution_clientToken :: Lens' StartChangeRequestExecution (Maybe Text)

The user-provided idempotency token. The token must be unique, is case insensitive, enforces the UUID format, and can't be reused.

startChangeRequestExecution_autoApprove :: Lens' StartChangeRequestExecution (Maybe Bool)

Indicates whether the change request can be approved automatically without the need for manual approvals.

If AutoApprovable is enabled in a change template, then setting AutoApprove to true in StartChangeRequestExecution creates a change request that bypasses approver review.

Change Calendar restrictions are not bypassed in this scenario. If the state of an associated calendar is CLOSED, change freeze approvers must still grant permission for this change request to run. If they don't, the change won't be processed until the calendar state is again OPEN.

startChangeRequestExecution_scheduledEndTime :: Lens' StartChangeRequestExecution (Maybe UTCTime)

The time that the requester expects the runbook workflow related to the change request to complete. The time is an estimate only that the requester provides for reviewers.

startChangeRequestExecution_parameters :: Lens' StartChangeRequestExecution (Maybe (HashMap Text [Text]))

A key-value map of parameters that match the declared parameters in the change template document.

startChangeRequestExecution_documentVersion :: Lens' StartChangeRequestExecution (Maybe Text)

The version of the change template document to run during the runbook workflow.

startChangeRequestExecution_changeRequestName :: Lens' StartChangeRequestExecution (Maybe Text)

The name of the change request associated with the runbook workflow to be run.

startChangeRequestExecution_tags :: Lens' StartChangeRequestExecution (Maybe [Tag])

Optional metadata that you assign to a resource. You can specify a maximum of five tags for a change request. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag a change request to identify an environment or target Amazon Web Services Region. In this case, you could specify the following key-value pairs:

Key=Environment,Value=Production

Key=Region,Value=us-east-2

startChangeRequestExecution_documentName :: Lens' StartChangeRequestExecution Text

The name of the change template document to run during the runbook workflow.

startChangeRequestExecution_runbooks :: Lens' StartChangeRequestExecution (NonEmpty Runbook)

Information about the Automation runbooks that are run during the runbook workflow.

The Automation runbooks specified for the runbook workflow can't run until all required approvals for the change request have been received.

Destructuring the Response

data StartChangeRequestExecutionResponse

See: newStartChangeRequestExecutionResponse smart constructor.

Constructors

StartChangeRequestExecutionResponse'

Fields

automationExecutionId :: Maybe Text

The unique ID of a runbook workflow operation. (A runbook workflow is a type of Automation operation.)

httpStatus :: Int

The response's http status code.

newStartChangeRequestExecutionResponse

Arguments

:: Int
$sel:httpStatus:StartChangeRequestExecutionResponse'

-> StartChangeRequestExecutionResponse

Create a value of StartChangeRequestExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:automationExecutionId:StartChangeRequestExecutionResponse', startChangeRequestExecutionResponse_automationExecutionId - The unique ID of a runbook workflow operation. (A runbook workflow is a type of Automation operation.)

$sel:httpStatus:StartChangeRequestExecutionResponse', startChangeRequestExecutionResponse_httpStatus - The response's http status code.

Response Lenses

startChangeRequestExecutionResponse_automationExecutionId :: Lens' StartChangeRequestExecutionResponse (Maybe Text)

The unique ID of a runbook workflow operation. (A runbook workflow is a type of Automation operation.)

startChangeRequestExecutionResponse_httpStatus :: Lens' StartChangeRequestExecutionResponse Int

The response's http status code.