Network.AWS.SSM.Types.AssociationDescription

Description

Documentation

data AssociationDescription

Describes the parameters for a document.

See: newAssociationDescription smart constructor.

Constructors

AssociationDescription'

Fields

maxErrors :: Maybe Text

The number of errors that are allowed before the system stops sending requests to run the association on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops sending requests when the fourth error is received. If you specify 0, then the system stops sending requests after the first error is returned. If you run an association on 50 instances and set MaxError to 10%, then the system stops sending the request when the sixth error is received.

Executions that are already running an association when MaxErrors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set MaxConcurrency to 1 so that executions proceed one at a time.

status :: Maybe AssociationStatus

The association status.

instanceId :: Maybe Text

The instance ID.

lastExecutionDate :: Maybe POSIX

The date on which the association was last run.

lastUpdateAssociationDate :: Maybe POSIX

The date when the association was last updated.

complianceSeverity :: Maybe AssociationComplianceSeverity

The severity level that is assigned to the association.

overview :: Maybe AssociationOverview

Information about the association.

automationTargetParameterName :: Maybe Text

Specify the target for the association. This target is required for associations that use an Automation runbook and target resources by using rate controls. Automation is a capability of Amazon Web Services Systems Manager.

targets :: Maybe [Target]

The instances targeted by the request.

name :: Maybe Text

The name of the SSM document.

scheduleExpression :: Maybe Text

A cron expression that specifies a schedule when the association runs.

targetLocations :: Maybe (NonEmpty TargetLocation)

The combination of Amazon Web Services Regions and Amazon Web Services accounts where you want to run the association.

date :: Maybe POSIX

The date when the association was made.

associationId :: Maybe Text

The association ID.

maxConcurrency :: Maybe Text

The maximum number of targets allowed to run the association at the same time. You can specify a number, for example 10, or a percentage of the target set, for example 10%. The default value is 100%, which means all targets run the association at the same time.

If a new instance starts and attempts to run an association while Systems Manager is running MaxConcurrency associations, the association is allowed to run. During the next association interval, the new instance will process its association within the limit specified for MaxConcurrency.

calendarNames :: Maybe [Text]

The names or Amazon Resource Names (ARNs) of the Change Calendar type documents your associations are gated under. The associations only run when that change calendar is open. For more information, see Amazon Web Services Systems Manager Change Calendar.

associationName :: Maybe Text

The association name.

associationVersion :: Maybe Text

The association version.

lastSuccessfulExecutionDate :: Maybe POSIX

The last date on which the association was successfully run.

documentVersion :: Maybe Text

The document version.

parameters :: Maybe (HashMap Text [Text])

A description of the parameters for a document.

syncCompliance :: Maybe AssociationSyncCompliance

The mode for generating association compliance. You can specify AUTO or MANUAL. In AUTO mode, the system uses the status of the association execution to determine the compliance status. If the association execution runs successfully, then the association is COMPLIANT. If the association execution doesn't run successfully, the association is NON-COMPLIANT.

In MANUAL mode, you must specify the AssociationId as a parameter for the PutComplianceItems API operation. In this case, compliance data isn't managed by State Manager, a capability of Amazon Web Services Systems Manager. It is managed by your direct call to the PutComplianceItems API operation.

By default, all associations use AUTO mode.

outputLocation :: Maybe InstanceAssociationOutputLocation

An S3 bucket where you want to store the output details of the request.

applyOnlyAtCronInterval :: Maybe Bool

By default, when you create a new associations, the system runs it immediately after it is created and then according to the schedule you specified. Specify this option if you don't want an association to run immediately after you create it. This parameter isn't supported for rate expressions.

newAssociationDescription :: AssociationDescription

Create a value of AssociationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxErrors:AssociationDescription', associationDescription_maxErrors - The number of errors that are allowed before the system stops sending requests to run the association on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops sending requests when the fourth error is received. If you specify 0, then the system stops sending requests after the first error is returned. If you run an association on 50 instances and set MaxError to 10%, then the system stops sending the request when the sixth error is received.

Executions that are already running an association when MaxErrors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set MaxConcurrency to 1 so that executions proceed one at a time.

$sel:status:AssociationDescription', associationDescription_status - The association status.

$sel:instanceId:AssociationDescription', associationDescription_instanceId - The instance ID.

$sel:lastExecutionDate:AssociationDescription', associationDescription_lastExecutionDate - The date on which the association was last run.

$sel:lastUpdateAssociationDate:AssociationDescription', associationDescription_lastUpdateAssociationDate - The date when the association was last updated.

$sel:complianceSeverity:AssociationDescription', associationDescription_complianceSeverity - The severity level that is assigned to the association.

$sel:overview:AssociationDescription', associationDescription_overview - Information about the association.

$sel:automationTargetParameterName:AssociationDescription', associationDescription_automationTargetParameterName - Specify the target for the association. This target is required for associations that use an Automation runbook and target resources by using rate controls. Automation is a capability of Amazon Web Services Systems Manager.

$sel:targets:AssociationDescription', associationDescription_targets - The instances targeted by the request.

$sel:name:AssociationDescription', associationDescription_name - The name of the SSM document.

$sel:scheduleExpression:AssociationDescription', associationDescription_scheduleExpression - A cron expression that specifies a schedule when the association runs.

$sel:targetLocations:AssociationDescription', associationDescription_targetLocations - The combination of Amazon Web Services Regions and Amazon Web Services accounts where you want to run the association.

$sel:date:AssociationDescription', associationDescription_date - The date when the association was made.

$sel:associationId:AssociationDescription', associationDescription_associationId - The association ID.

$sel:maxConcurrency:AssociationDescription', associationDescription_maxConcurrency - The maximum number of targets allowed to run the association at the same time. You can specify a number, for example 10, or a percentage of the target set, for example 10%. The default value is 100%, which means all targets run the association at the same time.

If a new instance starts and attempts to run an association while Systems Manager is running MaxConcurrency associations, the association is allowed to run. During the next association interval, the new instance will process its association within the limit specified for MaxConcurrency.

$sel:calendarNames:AssociationDescription', associationDescription_calendarNames - The names or Amazon Resource Names (ARNs) of the Change Calendar type documents your associations are gated under. The associations only run when that change calendar is open. For more information, see Amazon Web Services Systems Manager Change Calendar.

$sel:associationName:AssociationDescription', associationDescription_associationName - The association name.

$sel:associationVersion:AssociationDescription', associationDescription_associationVersion - The association version.

$sel:lastSuccessfulExecutionDate:AssociationDescription', associationDescription_lastSuccessfulExecutionDate - The last date on which the association was successfully run.

$sel:documentVersion:AssociationDescription', associationDescription_documentVersion - The document version.

$sel:parameters:AssociationDescription', associationDescription_parameters - A description of the parameters for a document.

$sel:syncCompliance:AssociationDescription', associationDescription_syncCompliance - The mode for generating association compliance. You can specify AUTO or MANUAL. In AUTO mode, the system uses the status of the association execution to determine the compliance status. If the association execution runs successfully, then the association is COMPLIANT. If the association execution doesn't run successfully, the association is NON-COMPLIANT.

In MANUAL mode, you must specify the AssociationId as a parameter for the PutComplianceItems API operation. In this case, compliance data isn't managed by State Manager, a capability of Amazon Web Services Systems Manager. It is managed by your direct call to the PutComplianceItems API operation.

By default, all associations use AUTO mode.

$sel:outputLocation:AssociationDescription', associationDescription_outputLocation - An S3 bucket where you want to store the output details of the request.

$sel:applyOnlyAtCronInterval:AssociationDescription', associationDescription_applyOnlyAtCronInterval - By default, when you create a new associations, the system runs it immediately after it is created and then according to the schedule you specified. Specify this option if you don't want an association to run immediately after you create it. This parameter isn't supported for rate expressions.

associationDescription_maxErrors :: Lens' AssociationDescription (Maybe Text)

The number of errors that are allowed before the system stops sending requests to run the association on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops sending requests when the fourth error is received. If you specify 0, then the system stops sending requests after the first error is returned. If you run an association on 50 instances and set MaxError to 10%, then the system stops sending the request when the sixth error is received.

Executions that are already running an association when MaxErrors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set MaxConcurrency to 1 so that executions proceed one at a time.

associationDescription_status :: Lens' AssociationDescription (Maybe AssociationStatus)

The association status.

associationDescription_instanceId :: Lens' AssociationDescription (Maybe Text)

The instance ID.

associationDescription_lastExecutionDate :: Lens' AssociationDescription (Maybe UTCTime)

The date on which the association was last run.

associationDescription_lastUpdateAssociationDate :: Lens' AssociationDescription (Maybe UTCTime)

The date when the association was last updated.

associationDescription_complianceSeverity :: Lens' AssociationDescription (Maybe AssociationComplianceSeverity)

The severity level that is assigned to the association.

associationDescription_overview :: Lens' AssociationDescription (Maybe AssociationOverview)

Information about the association.

associationDescription_automationTargetParameterName :: Lens' AssociationDescription (Maybe Text)

Specify the target for the association. This target is required for associations that use an Automation runbook and target resources by using rate controls. Automation is a capability of Amazon Web Services Systems Manager.

associationDescription_targets :: Lens' AssociationDescription (Maybe [Target])

The instances targeted by the request.

associationDescription_name :: Lens' AssociationDescription (Maybe Text)

The name of the SSM document.

associationDescription_scheduleExpression :: Lens' AssociationDescription (Maybe Text)

A cron expression that specifies a schedule when the association runs.

associationDescription_targetLocations :: Lens' AssociationDescription (Maybe (NonEmpty TargetLocation))

The combination of Amazon Web Services Regions and Amazon Web Services accounts where you want to run the association.

associationDescription_date :: Lens' AssociationDescription (Maybe UTCTime)

The date when the association was made.

associationDescription_associationId :: Lens' AssociationDescription (Maybe Text)

The association ID.

associationDescription_maxConcurrency :: Lens' AssociationDescription (Maybe Text)

The maximum number of targets allowed to run the association at the same time. You can specify a number, for example 10, or a percentage of the target set, for example 10%. The default value is 100%, which means all targets run the association at the same time.

If a new instance starts and attempts to run an association while Systems Manager is running MaxConcurrency associations, the association is allowed to run. During the next association interval, the new instance will process its association within the limit specified for MaxConcurrency.

associationDescription_calendarNames :: Lens' AssociationDescription (Maybe [Text])

The names or Amazon Resource Names (ARNs) of the Change Calendar type documents your associations are gated under. The associations only run when that change calendar is open. For more information, see Amazon Web Services Systems Manager Change Calendar.

associationDescription_associationName :: Lens' AssociationDescription (Maybe Text)

The association name.

associationDescription_associationVersion :: Lens' AssociationDescription (Maybe Text)

The association version.

associationDescription_lastSuccessfulExecutionDate :: Lens' AssociationDescription (Maybe UTCTime)

The last date on which the association was successfully run.

associationDescription_documentVersion :: Lens' AssociationDescription (Maybe Text)

The document version.

associationDescription_parameters :: Lens' AssociationDescription (Maybe (HashMap Text [Text]))

A description of the parameters for a document.

associationDescription_syncCompliance :: Lens' AssociationDescription (Maybe AssociationSyncCompliance)

The mode for generating association compliance. You can specify AUTO or MANUAL. In AUTO mode, the system uses the status of the association execution to determine the compliance status. If the association execution runs successfully, then the association is COMPLIANT. If the association execution doesn't run successfully, the association is NON-COMPLIANT.

In MANUAL mode, you must specify the AssociationId as a parameter for the PutComplianceItems API operation. In this case, compliance data isn't managed by State Manager, a capability of Amazon Web Services Systems Manager. It is managed by your direct call to the PutComplianceItems API operation.

By default, all associations use AUTO mode.

associationDescription_outputLocation :: Lens' AssociationDescription (Maybe InstanceAssociationOutputLocation)

An S3 bucket where you want to store the output details of the request.

associationDescription_applyOnlyAtCronInterval :: Lens' AssociationDescription (Maybe Bool)

By default, when you create a new associations, the system runs it immediately after it is created and then according to the schedule you specified. Specify this option if you don't want an association to run immediately after you create it. This parameter isn't supported for rate expressions.