Network.AWS.SSM.Types.AssociationExecution

Description

Documentation

data AssociationExecution

Includes information about the specified association.

See: newAssociationExecution smart constructor.

Constructors

AssociationExecution'

Fields

associationId :: Maybe Text

The association ID.

detailedStatus :: Maybe Text

Detailed status information about the execution.

status :: Maybe Text

The status of the association execution.

executionId :: Maybe Text

The execution ID for the association.

createdTime :: Maybe POSIX

The time the execution started.

resourceCountByStatus :: Maybe Text

An aggregate status of the resources in the execution based on the status type.

lastExecutionDate :: Maybe POSIX

The date of the last execution.

associationVersion :: Maybe Text

The association version.

newAssociationExecution :: AssociationExecution

Create a value of AssociationExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associationId:AssociationExecution', associationExecution_associationId - The association ID.

$sel:detailedStatus:AssociationExecution', associationExecution_detailedStatus - Detailed status information about the execution.

$sel:status:AssociationExecution', associationExecution_status - The status of the association execution.

$sel:executionId:AssociationExecution', associationExecution_executionId - The execution ID for the association.

$sel:createdTime:AssociationExecution', associationExecution_createdTime - The time the execution started.

$sel:resourceCountByStatus:AssociationExecution', associationExecution_resourceCountByStatus - An aggregate status of the resources in the execution based on the status type.

$sel:lastExecutionDate:AssociationExecution', associationExecution_lastExecutionDate - The date of the last execution.

$sel:associationVersion:AssociationExecution', associationExecution_associationVersion - The association version.

associationExecution_associationId :: Lens' AssociationExecution (Maybe Text)

The association ID.

associationExecution_detailedStatus :: Lens' AssociationExecution (Maybe Text)

Detailed status information about the execution.

associationExecution_status :: Lens' AssociationExecution (Maybe Text)

The status of the association execution.

associationExecution_executionId :: Lens' AssociationExecution (Maybe Text)

The execution ID for the association.

associationExecution_createdTime :: Lens' AssociationExecution (Maybe UTCTime)

The time the execution started.

associationExecution_resourceCountByStatus :: Lens' AssociationExecution (Maybe Text)

An aggregate status of the resources in the execution based on the status type.

associationExecution_lastExecutionDate :: Lens' AssociationExecution (Maybe UTCTime)

The date of the last execution.

associationExecution_associationVersion :: Lens' AssociationExecution (Maybe Text)

The association version.