Network.AWS.SSM.Types.AssociationOverview

Description

Documentation

data AssociationOverview

Information about the association.

See: newAssociationOverview smart constructor.

Constructors

AssociationOverview'

Fields

status :: Maybe Text

The status of the association. Status can be: Pending, Success, or Failed.

detailedStatus :: Maybe Text

A detailed status of the association.

associationStatusAggregatedCount :: Maybe (HashMap Text Int)

Returns the number of targets for the association status. For example, if you created an association with two instances, and one of them was successful, this would return the count of instances by status.

newAssociationOverview :: AssociationOverview

Create a value of AssociationOverview with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AssociationOverview', associationOverview_status - The status of the association. Status can be: Pending, Success, or Failed.

$sel:detailedStatus:AssociationOverview', associationOverview_detailedStatus - A detailed status of the association.

$sel:associationStatusAggregatedCount:AssociationOverview', associationOverview_associationStatusAggregatedCount - Returns the number of targets for the association status. For example, if you created an association with two instances, and one of them was successful, this would return the count of instances by status.

associationOverview_status :: Lens' AssociationOverview (Maybe Text)

The status of the association. Status can be: Pending, Success, or Failed.

associationOverview_detailedStatus :: Lens' AssociationOverview (Maybe Text)

A detailed status of the association.

associationOverview_associationStatusAggregatedCount :: Lens' AssociationOverview (Maybe (HashMap Text Int))

Returns the number of targets for the association status. For example, if you created an association with two instances, and one of them was successful, this would return the count of instances by status.