Network.AWS.SSM.Types.AssociationVersionInfo

Description

Documentation

data AssociationVersionInfo

Information about the association version.

See: newAssociationVersionInfo smart constructor.

Constructors

AssociationVersionInfo'

Fields

maxErrors :: Maybe Text

The number of errors that are allowed before the system stops sending requests to run the association on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops sending requests when the fourth error is received. If you specify 0, then the system stops sending requests after the first error is returned. If you run an association on 50 instances and set MaxError to 10%, then the system stops sending the request when the sixth error is received.

Executions that are already running an association when MaxErrors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set MaxConcurrency to 1 so that executions proceed one at a time.

createdDate :: Maybe POSIX

The date the association version was created.

complianceSeverity :: Maybe AssociationComplianceSeverity

The severity level that is assigned to the association.

targets :: Maybe [Target]

The targets specified for the association when the association version was created.

name :: Maybe Text

The name specified when the association was created.

scheduleExpression :: Maybe Text

The cron or rate schedule specified for the association when the association version was created.

targetLocations :: Maybe (NonEmpty TargetLocation)

The combination of Amazon Web Services Regions and Amazon Web Services accounts where you wanted to run the association when this association version was created.

associationId :: Maybe Text

The ID created by the system when the association was created.

maxConcurrency :: Maybe Text

The maximum number of targets allowed to run the association at the same time. You can specify a number, for example 10, or a percentage of the target set, for example 10%. The default value is 100%, which means all targets run the association at the same time.

If a new instance starts and attempts to run an association while Systems Manager is running MaxConcurrency associations, the association is allowed to run. During the next association interval, the new instance will process its association within the limit specified for MaxConcurrency.

calendarNames :: Maybe [Text]

The names or Amazon Resource Names (ARNs) of the Change Calendar type documents your associations are gated under. The associations for this version only run when that Change Calendar is open. For more information, see Amazon Web Services Systems Manager Change Calendar.

associationName :: Maybe Text

The name specified for the association version when the association version was created.

associationVersion :: Maybe Text

The association version.

documentVersion :: Maybe Text

The version of an Amazon Web Services Systems Manager document (SSM document) used when the association version was created.

parameters :: Maybe (HashMap Text [Text])

Parameters specified when the association version was created.

syncCompliance :: Maybe AssociationSyncCompliance

The mode for generating association compliance. You can specify AUTO or MANUAL. In AUTO mode, the system uses the status of the association execution to determine the compliance status. If the association execution runs successfully, then the association is COMPLIANT. If the association execution doesn't run successfully, the association is NON-COMPLIANT.

In MANUAL mode, you must specify the AssociationId as a parameter for the PutComplianceItems API operation. In this case, compliance data isn't managed by State Manager, a capability of Amazon Web Services Systems Manager. It is managed by your direct call to the PutComplianceItems API operation.

By default, all associations use AUTO mode.

outputLocation :: Maybe InstanceAssociationOutputLocation

The location in Amazon S3 specified for the association when the association version was created.

applyOnlyAtCronInterval :: Maybe Bool

By default, when you create a new associations, the system runs it immediately after it is created and then according to the schedule you specified. Specify this option if you don't want an association to run immediately after you create it. This parameter isn't supported for rate expressions.

newAssociationVersionInfo :: AssociationVersionInfo

Create a value of AssociationVersionInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxErrors:AssociationVersionInfo', associationVersionInfo_maxErrors - The number of errors that are allowed before the system stops sending requests to run the association on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops sending requests when the fourth error is received. If you specify 0, then the system stops sending requests after the first error is returned. If you run an association on 50 instances and set MaxError to 10%, then the system stops sending the request when the sixth error is received.

Executions that are already running an association when MaxErrors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set MaxConcurrency to 1 so that executions proceed one at a time.

$sel:createdDate:AssociationVersionInfo', associationVersionInfo_createdDate - The date the association version was created.

$sel:complianceSeverity:AssociationVersionInfo', associationVersionInfo_complianceSeverity - The severity level that is assigned to the association.

$sel:targets:AssociationVersionInfo', associationVersionInfo_targets - The targets specified for the association when the association version was created.

$sel:name:AssociationVersionInfo', associationVersionInfo_name - The name specified when the association was created.

$sel:scheduleExpression:AssociationVersionInfo', associationVersionInfo_scheduleExpression - The cron or rate schedule specified for the association when the association version was created.

$sel:targetLocations:AssociationVersionInfo', associationVersionInfo_targetLocations - The combination of Amazon Web Services Regions and Amazon Web Services accounts where you wanted to run the association when this association version was created.

$sel:associationId:AssociationVersionInfo', associationVersionInfo_associationId - The ID created by the system when the association was created.

$sel:maxConcurrency:AssociationVersionInfo', associationVersionInfo_maxConcurrency - The maximum number of targets allowed to run the association at the same time. You can specify a number, for example 10, or a percentage of the target set, for example 10%. The default value is 100%, which means all targets run the association at the same time.

If a new instance starts and attempts to run an association while Systems Manager is running MaxConcurrency associations, the association is allowed to run. During the next association interval, the new instance will process its association within the limit specified for MaxConcurrency.

$sel:calendarNames:AssociationVersionInfo', associationVersionInfo_calendarNames - The names or Amazon Resource Names (ARNs) of the Change Calendar type documents your associations are gated under. The associations for this version only run when that Change Calendar is open. For more information, see Amazon Web Services Systems Manager Change Calendar.

$sel:associationName:AssociationVersionInfo', associationVersionInfo_associationName - The name specified for the association version when the association version was created.

$sel:associationVersion:AssociationVersionInfo', associationVersionInfo_associationVersion - The association version.

$sel:documentVersion:AssociationVersionInfo', associationVersionInfo_documentVersion - The version of an Amazon Web Services Systems Manager document (SSM document) used when the association version was created.

$sel:parameters:AssociationVersionInfo', associationVersionInfo_parameters - Parameters specified when the association version was created.

$sel:syncCompliance:AssociationVersionInfo', associationVersionInfo_syncCompliance - The mode for generating association compliance. You can specify AUTO or MANUAL. In AUTO mode, the system uses the status of the association execution to determine the compliance status. If the association execution runs successfully, then the association is COMPLIANT. If the association execution doesn't run successfully, the association is NON-COMPLIANT.

In MANUAL mode, you must specify the AssociationId as a parameter for the PutComplianceItems API operation. In this case, compliance data isn't managed by State Manager, a capability of Amazon Web Services Systems Manager. It is managed by your direct call to the PutComplianceItems API operation.

By default, all associations use AUTO mode.

$sel:outputLocation:AssociationVersionInfo', associationVersionInfo_outputLocation - The location in Amazon S3 specified for the association when the association version was created.

$sel:applyOnlyAtCronInterval:AssociationVersionInfo', associationVersionInfo_applyOnlyAtCronInterval - By default, when you create a new associations, the system runs it immediately after it is created and then according to the schedule you specified. Specify this option if you don't want an association to run immediately after you create it. This parameter isn't supported for rate expressions.

associationVersionInfo_maxErrors :: Lens' AssociationVersionInfo (Maybe Text)

The number of errors that are allowed before the system stops sending requests to run the association on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops sending requests when the fourth error is received. If you specify 0, then the system stops sending requests after the first error is returned. If you run an association on 50 instances and set MaxError to 10%, then the system stops sending the request when the sixth error is received.

Executions that are already running an association when MaxErrors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set MaxConcurrency to 1 so that executions proceed one at a time.

associationVersionInfo_createdDate :: Lens' AssociationVersionInfo (Maybe UTCTime)

The date the association version was created.

associationVersionInfo_complianceSeverity :: Lens' AssociationVersionInfo (Maybe AssociationComplianceSeverity)

The severity level that is assigned to the association.

associationVersionInfo_targets :: Lens' AssociationVersionInfo (Maybe [Target])

The targets specified for the association when the association version was created.

associationVersionInfo_name :: Lens' AssociationVersionInfo (Maybe Text)

The name specified when the association was created.

associationVersionInfo_scheduleExpression :: Lens' AssociationVersionInfo (Maybe Text)

The cron or rate schedule specified for the association when the association version was created.

associationVersionInfo_targetLocations :: Lens' AssociationVersionInfo (Maybe (NonEmpty TargetLocation))

The combination of Amazon Web Services Regions and Amazon Web Services accounts where you wanted to run the association when this association version was created.

associationVersionInfo_associationId :: Lens' AssociationVersionInfo (Maybe Text)

The ID created by the system when the association was created.

associationVersionInfo_maxConcurrency :: Lens' AssociationVersionInfo (Maybe Text)

The maximum number of targets allowed to run the association at the same time. You can specify a number, for example 10, or a percentage of the target set, for example 10%. The default value is 100%, which means all targets run the association at the same time.

If a new instance starts and attempts to run an association while Systems Manager is running MaxConcurrency associations, the association is allowed to run. During the next association interval, the new instance will process its association within the limit specified for MaxConcurrency.

associationVersionInfo_calendarNames :: Lens' AssociationVersionInfo (Maybe [Text])

The names or Amazon Resource Names (ARNs) of the Change Calendar type documents your associations are gated under. The associations for this version only run when that Change Calendar is open. For more information, see Amazon Web Services Systems Manager Change Calendar.

associationVersionInfo_associationName :: Lens' AssociationVersionInfo (Maybe Text)

The name specified for the association version when the association version was created.

associationVersionInfo_associationVersion :: Lens' AssociationVersionInfo (Maybe Text)

The association version.

associationVersionInfo_documentVersion :: Lens' AssociationVersionInfo (Maybe Text)

The version of an Amazon Web Services Systems Manager document (SSM document) used when the association version was created.

associationVersionInfo_parameters :: Lens' AssociationVersionInfo (Maybe (HashMap Text [Text]))

Parameters specified when the association version was created.

associationVersionInfo_syncCompliance :: Lens' AssociationVersionInfo (Maybe AssociationSyncCompliance)

The mode for generating association compliance. You can specify AUTO or MANUAL. In AUTO mode, the system uses the status of the association execution to determine the compliance status. If the association execution runs successfully, then the association is COMPLIANT. If the association execution doesn't run successfully, the association is NON-COMPLIANT.

In MANUAL mode, you must specify the AssociationId as a parameter for the PutComplianceItems API operation. In this case, compliance data isn't managed by State Manager, a capability of Amazon Web Services Systems Manager. It is managed by your direct call to the PutComplianceItems API operation.

By default, all associations use AUTO mode.

associationVersionInfo_outputLocation :: Lens' AssociationVersionInfo (Maybe InstanceAssociationOutputLocation)

The location in Amazon S3 specified for the association when the association version was created.

associationVersionInfo_applyOnlyAtCronInterval :: Lens' AssociationVersionInfo (Maybe Bool)

By default, when you create a new associations, the system runs it immediately after it is created and then according to the schedule you specified. Specify this option if you don't want an association to run immediately after you create it. This parameter isn't supported for rate expressions.