Network.AWS.SSM.Types.AttachmentContent

Description

Documentation

data AttachmentContent

A structure that includes attributes that describe a document attachment.

See: newAttachmentContent smart constructor.

Constructors

AttachmentContent'

Fields

hash :: Maybe Text

The cryptographic hash value of the document content.

size :: Maybe Integer

The size of an attachment in bytes.

url :: Maybe Text

The URL location of the attachment content.

name :: Maybe Text

The name of an attachment.

hashType :: Maybe AttachmentHashType

The hash algorithm used to calculate the hash value.

newAttachmentContent :: AttachmentContent

Create a value of AttachmentContent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hash:AttachmentContent', attachmentContent_hash - The cryptographic hash value of the document content.

$sel:size:AttachmentContent', attachmentContent_size - The size of an attachment in bytes.

$sel:url:AttachmentContent', attachmentContent_url - The URL location of the attachment content.

$sel:name:AttachmentContent', attachmentContent_name - The name of an attachment.

$sel:hashType:AttachmentContent', attachmentContent_hashType - The hash algorithm used to calculate the hash value.

attachmentContent_hash :: Lens' AttachmentContent (Maybe Text)

The cryptographic hash value of the document content.

attachmentContent_size :: Lens' AttachmentContent (Maybe Integer)

The size of an attachment in bytes.

attachmentContent_url :: Lens' AttachmentContent (Maybe Text)

The URL location of the attachment content.

attachmentContent_name :: Lens' AttachmentContent (Maybe Text)

The name of an attachment.

attachmentContent_hashType :: Lens' AttachmentContent (Maybe AttachmentHashType)

The hash algorithm used to calculate the hash value.