Network.AWS.SSM.Types.AutomationExecutionMetadata

Description

Documentation

data AutomationExecutionMetadata

Details about a specific Automation execution.

See: newAutomationExecutionMetadata smart constructor.

Constructors

AutomationExecutionMetadata'

Fields

maxErrors :: Maybe Text

The MaxErrors value specified by the user when starting the automation.

currentAction :: Maybe Text

The action of the step that is currently running.

parentAutomationExecutionId :: Maybe Text

The execution ID of the parent automation.

outputs :: Maybe (HashMap Text [Text])

The list of execution outputs as defined in the Automation runbook.

failureMessage :: Maybe Text

The list of execution outputs as defined in the Automation runbook.

mode :: Maybe ExecutionMode

The Automation execution mode.

executionEndTime :: Maybe POSIX

The time the execution finished. This isn't populated if the execution is still in progress.

changeRequestName :: Maybe Text

The name of the Change Manager change request.

automationExecutionId :: Maybe Text

The execution ID.

executedBy :: Maybe Text

The IAM role ARN of the user who ran the automation.

documentName :: Maybe Text

The name of the Automation runbook used during execution.

targets :: Maybe [Target]

The targets defined by the user when starting the automation.

resolvedTargets :: Maybe ResolvedTargets

A list of targets that resolved during the execution.

executionStartTime :: Maybe POSIX

The time the execution started.

automationType :: Maybe AutomationType

Use this filter with DescribeAutomationExecutions. Specify either Local or CrossAccount. CrossAccount is an Automation that runs in multiple Amazon Web Services Regions and Amazon Web Services accounts. For more information, see Running Automation workflows in multiple Amazon Web Services Regions and accounts in the Amazon Web Services Systems Manager User Guide.

targetParameterName :: Maybe Text

The list of execution outputs as defined in the Automation runbook.

currentStepName :: Maybe Text

The name of the step that is currently running.

opsItemId :: Maybe Text

The ID of an OpsItem that is created to represent a Change Manager change request.

associationId :: Maybe Text

The ID of a State Manager association used in the Automation operation.

scheduledTime :: Maybe POSIX

The date and time the Automation operation is scheduled to start.

maxConcurrency :: Maybe Text

The MaxConcurrency value specified by the user when starting the automation.

target :: Maybe Text

The list of execution outputs as defined in the Automation runbook.

automationExecutionStatus :: Maybe AutomationExecutionStatus

The status of the execution.

targetMaps :: Maybe [HashMap Text [Text]]

The specified key-value mapping of document parameters to target resources.

runbooks :: Maybe (NonEmpty Runbook)

Information about the Automation runbooks that are run during a runbook workflow in Change Manager.

The Automation runbooks specified for the runbook workflow can't run until all required approvals for the change request have been received.

automationSubtype :: Maybe AutomationSubtype

The subtype of the Automation operation. Currently, the only supported value is ChangeRequest.

documentVersion :: Maybe Text

The document version used during the execution.

logFile :: Maybe Text

An S3 bucket where execution information is stored.

newAutomationExecutionMetadata :: AutomationExecutionMetadata

Create a value of AutomationExecutionMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxErrors:AutomationExecutionMetadata', automationExecutionMetadata_maxErrors - The MaxErrors value specified by the user when starting the automation.

$sel:currentAction:AutomationExecutionMetadata', automationExecutionMetadata_currentAction - The action of the step that is currently running.

$sel:parentAutomationExecutionId:AutomationExecutionMetadata', automationExecutionMetadata_parentAutomationExecutionId - The execution ID of the parent automation.

$sel:outputs:AutomationExecutionMetadata', automationExecutionMetadata_outputs - The list of execution outputs as defined in the Automation runbook.

$sel:failureMessage:AutomationExecutionMetadata', automationExecutionMetadata_failureMessage - The list of execution outputs as defined in the Automation runbook.

$sel:mode:AutomationExecutionMetadata', automationExecutionMetadata_mode - The Automation execution mode.

$sel:executionEndTime:AutomationExecutionMetadata', automationExecutionMetadata_executionEndTime - The time the execution finished. This isn't populated if the execution is still in progress.

$sel:changeRequestName:AutomationExecutionMetadata', automationExecutionMetadata_changeRequestName - The name of the Change Manager change request.

$sel:automationExecutionId:AutomationExecutionMetadata', automationExecutionMetadata_automationExecutionId - The execution ID.

$sel:executedBy:AutomationExecutionMetadata', automationExecutionMetadata_executedBy - The IAM role ARN of the user who ran the automation.

$sel:documentName:AutomationExecutionMetadata', automationExecutionMetadata_documentName - The name of the Automation runbook used during execution.

$sel:targets:AutomationExecutionMetadata', automationExecutionMetadata_targets - The targets defined by the user when starting the automation.

$sel:resolvedTargets:AutomationExecutionMetadata', automationExecutionMetadata_resolvedTargets - A list of targets that resolved during the execution.

$sel:executionStartTime:AutomationExecutionMetadata', automationExecutionMetadata_executionStartTime - The time the execution started.

$sel:automationType:AutomationExecutionMetadata', automationExecutionMetadata_automationType - Use this filter with DescribeAutomationExecutions. Specify either Local or CrossAccount. CrossAccount is an Automation that runs in multiple Amazon Web Services Regions and Amazon Web Services accounts. For more information, see Running Automation workflows in multiple Amazon Web Services Regions and accounts in the Amazon Web Services Systems Manager User Guide.

$sel:targetParameterName:AutomationExecutionMetadata', automationExecutionMetadata_targetParameterName - The list of execution outputs as defined in the Automation runbook.

$sel:currentStepName:AutomationExecutionMetadata', automationExecutionMetadata_currentStepName - The name of the step that is currently running.

$sel:opsItemId:AutomationExecutionMetadata', automationExecutionMetadata_opsItemId - The ID of an OpsItem that is created to represent a Change Manager change request.

$sel:associationId:AutomationExecutionMetadata', automationExecutionMetadata_associationId - The ID of a State Manager association used in the Automation operation.

$sel:scheduledTime:AutomationExecutionMetadata', automationExecutionMetadata_scheduledTime - The date and time the Automation operation is scheduled to start.

$sel:maxConcurrency:AutomationExecutionMetadata', automationExecutionMetadata_maxConcurrency - The MaxConcurrency value specified by the user when starting the automation.

$sel:target:AutomationExecutionMetadata', automationExecutionMetadata_target - The list of execution outputs as defined in the Automation runbook.

$sel:automationExecutionStatus:AutomationExecutionMetadata', automationExecutionMetadata_automationExecutionStatus - The status of the execution.

$sel:targetMaps:AutomationExecutionMetadata', automationExecutionMetadata_targetMaps - The specified key-value mapping of document parameters to target resources.

$sel:runbooks:AutomationExecutionMetadata', automationExecutionMetadata_runbooks - Information about the Automation runbooks that are run during a runbook workflow in Change Manager.

The Automation runbooks specified for the runbook workflow can't run until all required approvals for the change request have been received.

$sel:automationSubtype:AutomationExecutionMetadata', automationExecutionMetadata_automationSubtype - The subtype of the Automation operation. Currently, the only supported value is ChangeRequest.

$sel:documentVersion:AutomationExecutionMetadata', automationExecutionMetadata_documentVersion - The document version used during the execution.

$sel:logFile:AutomationExecutionMetadata', automationExecutionMetadata_logFile - An S3 bucket where execution information is stored.

automationExecutionMetadata_maxErrors :: Lens' AutomationExecutionMetadata (Maybe Text)

The MaxErrors value specified by the user when starting the automation.

automationExecutionMetadata_currentAction :: Lens' AutomationExecutionMetadata (Maybe Text)

The action of the step that is currently running.

automationExecutionMetadata_parentAutomationExecutionId :: Lens' AutomationExecutionMetadata (Maybe Text)

The execution ID of the parent automation.

automationExecutionMetadata_outputs :: Lens' AutomationExecutionMetadata (Maybe (HashMap Text [Text]))

The list of execution outputs as defined in the Automation runbook.

automationExecutionMetadata_failureMessage :: Lens' AutomationExecutionMetadata (Maybe Text)

The list of execution outputs as defined in the Automation runbook.

automationExecutionMetadata_mode :: Lens' AutomationExecutionMetadata (Maybe ExecutionMode)

The Automation execution mode.

automationExecutionMetadata_executionEndTime :: Lens' AutomationExecutionMetadata (Maybe UTCTime)

The time the execution finished. This isn't populated if the execution is still in progress.

automationExecutionMetadata_changeRequestName :: Lens' AutomationExecutionMetadata (Maybe Text)

The name of the Change Manager change request.

automationExecutionMetadata_automationExecutionId :: Lens' AutomationExecutionMetadata (Maybe Text)

The execution ID.

automationExecutionMetadata_executedBy :: Lens' AutomationExecutionMetadata (Maybe Text)

The IAM role ARN of the user who ran the automation.

automationExecutionMetadata_documentName :: Lens' AutomationExecutionMetadata (Maybe Text)

The name of the Automation runbook used during execution.

automationExecutionMetadata_targets :: Lens' AutomationExecutionMetadata (Maybe [Target])

The targets defined by the user when starting the automation.

automationExecutionMetadata_resolvedTargets :: Lens' AutomationExecutionMetadata (Maybe ResolvedTargets)

A list of targets that resolved during the execution.

automationExecutionMetadata_executionStartTime :: Lens' AutomationExecutionMetadata (Maybe UTCTime)

The time the execution started.

automationExecutionMetadata_automationType :: Lens' AutomationExecutionMetadata (Maybe AutomationType)

Use this filter with DescribeAutomationExecutions. Specify either Local or CrossAccount. CrossAccount is an Automation that runs in multiple Amazon Web Services Regions and Amazon Web Services accounts. For more information, see Running Automation workflows in multiple Amazon Web Services Regions and accounts in the Amazon Web Services Systems Manager User Guide.

automationExecutionMetadata_targetParameterName :: Lens' AutomationExecutionMetadata (Maybe Text)

The list of execution outputs as defined in the Automation runbook.

automationExecutionMetadata_currentStepName :: Lens' AutomationExecutionMetadata (Maybe Text)

The name of the step that is currently running.

automationExecutionMetadata_opsItemId :: Lens' AutomationExecutionMetadata (Maybe Text)

The ID of an OpsItem that is created to represent a Change Manager change request.

automationExecutionMetadata_associationId :: Lens' AutomationExecutionMetadata (Maybe Text)

The ID of a State Manager association used in the Automation operation.

automationExecutionMetadata_scheduledTime :: Lens' AutomationExecutionMetadata (Maybe UTCTime)

The date and time the Automation operation is scheduled to start.

automationExecutionMetadata_maxConcurrency :: Lens' AutomationExecutionMetadata (Maybe Text)

The MaxConcurrency value specified by the user when starting the automation.

automationExecutionMetadata_target :: Lens' AutomationExecutionMetadata (Maybe Text)

The list of execution outputs as defined in the Automation runbook.

automationExecutionMetadata_automationExecutionStatus :: Lens' AutomationExecutionMetadata (Maybe AutomationExecutionStatus)

The status of the execution.

automationExecutionMetadata_targetMaps :: Lens' AutomationExecutionMetadata (Maybe [HashMap Text [Text]])

The specified key-value mapping of document parameters to target resources.

automationExecutionMetadata_runbooks :: Lens' AutomationExecutionMetadata (Maybe (NonEmpty Runbook))

Information about the Automation runbooks that are run during a runbook workflow in Change Manager.

The Automation runbooks specified for the runbook workflow can't run until all required approvals for the change request have been received.

automationExecutionMetadata_automationSubtype :: Lens' AutomationExecutionMetadata (Maybe AutomationSubtype)

The subtype of the Automation operation. Currently, the only supported value is ChangeRequest.

automationExecutionMetadata_documentVersion :: Lens' AutomationExecutionMetadata (Maybe Text)

The document version used during the execution.

automationExecutionMetadata_logFile :: Lens' AutomationExecutionMetadata (Maybe Text)

An S3 bucket where execution information is stored.