Network.AWS.SSM.Types.BaselineOverride

Description

Documentation

data BaselineOverride

Defines the basic information about a patch baseline override.

See: newBaselineOverride smart constructor.

Constructors

BaselineOverride'

Fields

rejectedPatches :: Maybe [Text]

A list of explicitly rejected patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

sources :: Maybe [PatchSource]

Information about the patches to use to update the instances, including target operating systems and source repositories. Applies to Linux instances only.

approvedPatchesEnableNonSecurity :: Maybe Bool

Indicates whether the list of approved patches includes non-security updates that should be applied to the instances. The default value is false. Applies to Linux instances only.

approvedPatchesComplianceLevel :: Maybe PatchComplianceLevel

Defines the compliance level for approved patches. When an approved patch is reported as missing, this value describes the severity of the compliance violation.

approvedPatches :: Maybe [Text]

A list of explicitly approved patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

rejectedPatchesAction :: Maybe PatchAction

The action for Patch Manager to take on patches included in the RejectedPackages list. A patch can be allowed only if it is a dependency of another package, or blocked entirely along with packages that include it as a dependency.

operatingSystem :: Maybe OperatingSystem

The operating system rule used by the patch baseline override.

globalFilters :: Maybe PatchFilterGroup

approvalRules :: Maybe PatchRuleGroup

newBaselineOverride :: BaselineOverride

Create a value of BaselineOverride with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rejectedPatches:BaselineOverride', baselineOverride_rejectedPatches - A list of explicitly rejected patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

$sel:sources:BaselineOverride', baselineOverride_sources - Information about the patches to use to update the instances, including target operating systems and source repositories. Applies to Linux instances only.

$sel:approvedPatchesEnableNonSecurity:BaselineOverride', baselineOverride_approvedPatchesEnableNonSecurity - Indicates whether the list of approved patches includes non-security updates that should be applied to the instances. The default value is false. Applies to Linux instances only.

$sel:approvedPatchesComplianceLevel:BaselineOverride', baselineOverride_approvedPatchesComplianceLevel - Defines the compliance level for approved patches. When an approved patch is reported as missing, this value describes the severity of the compliance violation.

$sel:approvedPatches:BaselineOverride', baselineOverride_approvedPatches - A list of explicitly approved patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

$sel:rejectedPatchesAction:BaselineOverride', baselineOverride_rejectedPatchesAction - The action for Patch Manager to take on patches included in the RejectedPackages list. A patch can be allowed only if it is a dependency of another package, or blocked entirely along with packages that include it as a dependency.

$sel:operatingSystem:BaselineOverride', baselineOverride_operatingSystem - The operating system rule used by the patch baseline override.

$sel:globalFilters:BaselineOverride', baselineOverride_globalFilters - Undocumented member.

$sel:approvalRules:BaselineOverride', baselineOverride_approvalRules - Undocumented member.

baselineOverride_rejectedPatches :: Lens' BaselineOverride (Maybe [Text])

A list of explicitly rejected patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

baselineOverride_sources :: Lens' BaselineOverride (Maybe [PatchSource])

Information about the patches to use to update the instances, including target operating systems and source repositories. Applies to Linux instances only.

baselineOverride_approvedPatchesEnableNonSecurity :: Lens' BaselineOverride (Maybe Bool)

Indicates whether the list of approved patches includes non-security updates that should be applied to the instances. The default value is false. Applies to Linux instances only.

baselineOverride_approvedPatchesComplianceLevel :: Lens' BaselineOverride (Maybe PatchComplianceLevel)

Defines the compliance level for approved patches. When an approved patch is reported as missing, this value describes the severity of the compliance violation.

baselineOverride_approvedPatches :: Lens' BaselineOverride (Maybe [Text])

A list of explicitly approved patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

baselineOverride_rejectedPatchesAction :: Lens' BaselineOverride (Maybe PatchAction)

The action for Patch Manager to take on patches included in the RejectedPackages list. A patch can be allowed only if it is a dependency of another package, or blocked entirely along with packages that include it as a dependency.

baselineOverride_operatingSystem :: Lens' BaselineOverride (Maybe OperatingSystem)

The operating system rule used by the patch baseline override.

baselineOverride_globalFilters :: Lens' BaselineOverride (Maybe PatchFilterGroup)

Undocumented member.

baselineOverride_approvalRules :: Lens' BaselineOverride (Maybe PatchRuleGroup)

Undocumented member.