Network.AWS.SSM.Types.Command

Description

Documentation

data Command

Describes a command request.

See: newCommand smart constructor.

Constructors

Command'

Fields

instanceIds :: Maybe [Text]

The instance IDs against which this command was requested.

maxErrors :: Maybe Text

The maximum number of errors allowed before the system stops sending the command to additional targets. You can specify a number of errors, such as 10, or a percentage or errors, such as 10%. The default value is 0. For more information about how to use MaxErrors, see Running commands using Systems Manager Run Command in the Amazon Web Services Systems Manager User Guide.

notificationConfig :: Maybe NotificationConfig

Configurations for sending notifications about command status changes.

status :: Maybe CommandStatus

The status of the command.

expiresAfter :: Maybe POSIX

If this time is reached and the command hasn't already started running, it won't run. Calculated based on the ExpiresAfter user input provided as part of the SendCommand API operation.

requestedDateTime :: Maybe POSIX

The date and time the command was requested.

serviceRole :: Maybe Text

The Identity and Access Management (IAM) service role that Run Command, a capability of Amazon Web Services Systems Manager, uses to act on your behalf when sending notifications about command status changes.

statusDetails :: Maybe Text

A detailed status of the command execution. StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Understanding command statuses in the Amazon Web Services Systems Manager User Guide. StatusDetails can be one of the following values:

Pending: The command hasn't been sent to any instances.

In Progress: The command has been sent to at least one instance but hasn't reached a final state on all instances.

Success: The command successfully ran on all invocations. This is a terminal state.

Delivery Timed Out: The value of MaxErrors or more command invocations shows a status of Delivery Timed Out. This is a terminal state.

Execution Timed Out: The value of MaxErrors or more command invocations shows a status of Execution Timed Out. This is a terminal state.

Failed: The value of MaxErrors or more command invocations shows a status of Failed. This is a terminal state.

Incomplete: The command was attempted on all instances and one or more invocations doesn't have a value of Success but not enough invocations failed for the status to be Failed. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Rate Exceeded: The number of instances targeted by the command exceeded the account limit for pending invocations. The system has canceled the command before running it on any instance. This is a terminal state.

completedCount :: Maybe Int

The number of targets for which the command invocation reached a terminal state. Terminal states include the following: Success, Failed, Execution Timed Out, Delivery Timed Out, Canceled, Terminated, or Undeliverable.

outputS3BucketName :: Maybe Text

The S3 bucket where the responses to the command executions should be stored. This was requested when issuing the command.

errorCount :: Maybe Int

The number of targets for which the status is Failed or Execution Timed Out.

comment :: Maybe Text

User-specified information about the command, such as a brief description of what the command should do.

documentName :: Maybe Text

The name of the document requested for execution.

commandId :: Maybe Text

A unique identifier for this command.

targets :: Maybe [Target]

An array of search criteria that targets instances using a Key,Value combination that you specify. Targets is required if you don't provide one or more instance IDs in the call.

outputS3Region :: Maybe Text

(Deprecated) You can no longer specify this parameter. The system ignores it. Instead, Systems Manager automatically determines the Amazon Web Services Region of the S3 bucket.

maxConcurrency :: Maybe Text

The maximum number of instances that are allowed to run the command at the same time. You can specify a number of instances, such as 10, or a percentage of instances, such as 10%. The default value is 50. For more information about how to use MaxConcurrency, see Running commands using Systems Manager Run Command in the Amazon Web Services Systems Manager User Guide.

deliveryTimedOutCount :: Maybe Int

The number of targets for which the status is Delivery Timed Out.

timeoutSeconds :: Maybe Natural

The TimeoutSeconds value specified for a command.

outputS3KeyPrefix :: Maybe Text

The S3 directory path inside the bucket where the responses to the command executions should be stored. This was requested when issuing the command.

cloudWatchOutputConfig :: Maybe CloudWatchOutputConfig

Amazon CloudWatch Logs information where you want Amazon Web Services Systems Manager to send the command output.

documentVersion :: Maybe Text

The Systems Manager document (SSM document) version.

targetCount :: Maybe Int

The number of targets for the command.

parameters :: Maybe (HashMap Text [Text])

The parameter values to be inserted in the document when running the command.

newCommand :: Command

Create a value of Command with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceIds:Command', command_instanceIds - The instance IDs against which this command was requested.

$sel:maxErrors:Command', command_maxErrors - The maximum number of errors allowed before the system stops sending the command to additional targets. You can specify a number of errors, such as 10, or a percentage or errors, such as 10%. The default value is 0. For more information about how to use MaxErrors, see Running commands using Systems Manager Run Command in the Amazon Web Services Systems Manager User Guide.

$sel:notificationConfig:Command', command_notificationConfig - Configurations for sending notifications about command status changes.

$sel:status:Command', command_status - The status of the command.

$sel:expiresAfter:Command', command_expiresAfter - If this time is reached and the command hasn't already started running, it won't run. Calculated based on the ExpiresAfter user input provided as part of the SendCommand API operation.

$sel:requestedDateTime:Command', command_requestedDateTime - The date and time the command was requested.

$sel:serviceRole:Command', command_serviceRole - The Identity and Access Management (IAM) service role that Run Command, a capability of Amazon Web Services Systems Manager, uses to act on your behalf when sending notifications about command status changes.

$sel:statusDetails:Command', command_statusDetails - A detailed status of the command execution. StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Understanding command statuses in the Amazon Web Services Systems Manager User Guide. StatusDetails can be one of the following values:

Pending: The command hasn't been sent to any instances.

In Progress: The command has been sent to at least one instance but hasn't reached a final state on all instances.

Success: The command successfully ran on all invocations. This is a terminal state.

Delivery Timed Out: The value of MaxErrors or more command invocations shows a status of Delivery Timed Out. This is a terminal state.

Execution Timed Out: The value of MaxErrors or more command invocations shows a status of Execution Timed Out. This is a terminal state.

Failed: The value of MaxErrors or more command invocations shows a status of Failed. This is a terminal state.

Incomplete: The command was attempted on all instances and one or more invocations doesn't have a value of Success but not enough invocations failed for the status to be Failed. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Rate Exceeded: The number of instances targeted by the command exceeded the account limit for pending invocations. The system has canceled the command before running it on any instance. This is a terminal state.

$sel:completedCount:Command', command_completedCount - The number of targets for which the command invocation reached a terminal state. Terminal states include the following: Success, Failed, Execution Timed Out, Delivery Timed Out, Canceled, Terminated, or Undeliverable.

$sel:outputS3BucketName:Command', command_outputS3BucketName - The S3 bucket where the responses to the command executions should be stored. This was requested when issuing the command.

$sel:errorCount:Command', command_errorCount - The number of targets for which the status is Failed or Execution Timed Out.

$sel:comment:Command', command_comment - User-specified information about the command, such as a brief description of what the command should do.

$sel:documentName:Command', command_documentName - The name of the document requested for execution.

$sel:commandId:Command', command_commandId - A unique identifier for this command.

$sel:targets:Command', command_targets - An array of search criteria that targets instances using a Key,Value combination that you specify. Targets is required if you don't provide one or more instance IDs in the call.

$sel:outputS3Region:Command', command_outputS3Region - (Deprecated) You can no longer specify this parameter. The system ignores it. Instead, Systems Manager automatically determines the Amazon Web Services Region of the S3 bucket.

$sel:maxConcurrency:Command', command_maxConcurrency - The maximum number of instances that are allowed to run the command at the same time. You can specify a number of instances, such as 10, or a percentage of instances, such as 10%. The default value is 50. For more information about how to use MaxConcurrency, see Running commands using Systems Manager Run Command in the Amazon Web Services Systems Manager User Guide.

$sel:deliveryTimedOutCount:Command', command_deliveryTimedOutCount - The number of targets for which the status is Delivery Timed Out.

$sel:timeoutSeconds:Command', command_timeoutSeconds - The TimeoutSeconds value specified for a command.

$sel:outputS3KeyPrefix:Command', command_outputS3KeyPrefix - The S3 directory path inside the bucket where the responses to the command executions should be stored. This was requested when issuing the command.

$sel:cloudWatchOutputConfig:Command', command_cloudWatchOutputConfig - Amazon CloudWatch Logs information where you want Amazon Web Services Systems Manager to send the command output.

$sel:documentVersion:Command', command_documentVersion - The Systems Manager document (SSM document) version.

$sel:targetCount:Command', command_targetCount - The number of targets for the command.

$sel:parameters:Command', command_parameters - The parameter values to be inserted in the document when running the command.

command_instanceIds :: Lens' Command (Maybe [Text])

The instance IDs against which this command was requested.

command_maxErrors :: Lens' Command (Maybe Text)

The maximum number of errors allowed before the system stops sending the command to additional targets. You can specify a number of errors, such as 10, or a percentage or errors, such as 10%. The default value is 0. For more information about how to use MaxErrors, see Running commands using Systems Manager Run Command in the Amazon Web Services Systems Manager User Guide.

command_notificationConfig :: Lens' Command (Maybe NotificationConfig)

Configurations for sending notifications about command status changes.

command_status :: Lens' Command (Maybe CommandStatus)

The status of the command.

command_expiresAfter :: Lens' Command (Maybe UTCTime)

If this time is reached and the command hasn't already started running, it won't run. Calculated based on the ExpiresAfter user input provided as part of the SendCommand API operation.

command_requestedDateTime :: Lens' Command (Maybe UTCTime)

The date and time the command was requested.

command_serviceRole :: Lens' Command (Maybe Text)

The Identity and Access Management (IAM) service role that Run Command, a capability of Amazon Web Services Systems Manager, uses to act on your behalf when sending notifications about command status changes.

command_statusDetails :: Lens' Command (Maybe Text)

A detailed status of the command execution. StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Understanding command statuses in the Amazon Web Services Systems Manager User Guide. StatusDetails can be one of the following values:

Pending: The command hasn't been sent to any instances.

In Progress: The command has been sent to at least one instance but hasn't reached a final state on all instances.

Success: The command successfully ran on all invocations. This is a terminal state.

Delivery Timed Out: The value of MaxErrors or more command invocations shows a status of Delivery Timed Out. This is a terminal state.

Execution Timed Out: The value of MaxErrors or more command invocations shows a status of Execution Timed Out. This is a terminal state.

Failed: The value of MaxErrors or more command invocations shows a status of Failed. This is a terminal state.

Incomplete: The command was attempted on all instances and one or more invocations doesn't have a value of Success but not enough invocations failed for the status to be Failed. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Rate Exceeded: The number of instances targeted by the command exceeded the account limit for pending invocations. The system has canceled the command before running it on any instance. This is a terminal state.

command_completedCount :: Lens' Command (Maybe Int)

The number of targets for which the command invocation reached a terminal state. Terminal states include the following: Success, Failed, Execution Timed Out, Delivery Timed Out, Canceled, Terminated, or Undeliverable.

command_outputS3BucketName :: Lens' Command (Maybe Text)

The S3 bucket where the responses to the command executions should be stored. This was requested when issuing the command.

command_errorCount :: Lens' Command (Maybe Int)

The number of targets for which the status is Failed or Execution Timed Out.

command_comment :: Lens' Command (Maybe Text)

User-specified information about the command, such as a brief description of what the command should do.

command_documentName :: Lens' Command (Maybe Text)

The name of the document requested for execution.

command_commandId :: Lens' Command (Maybe Text)

A unique identifier for this command.

command_targets :: Lens' Command (Maybe [Target])

An array of search criteria that targets instances using a Key,Value combination that you specify. Targets is required if you don't provide one or more instance IDs in the call.

command_outputS3Region :: Lens' Command (Maybe Text)

(Deprecated) You can no longer specify this parameter. The system ignores it. Instead, Systems Manager automatically determines the Amazon Web Services Region of the S3 bucket.

command_maxConcurrency :: Lens' Command (Maybe Text)

The maximum number of instances that are allowed to run the command at the same time. You can specify a number of instances, such as 10, or a percentage of instances, such as 10%. The default value is 50. For more information about how to use MaxConcurrency, see Running commands using Systems Manager Run Command in the Amazon Web Services Systems Manager User Guide.

command_deliveryTimedOutCount :: Lens' Command (Maybe Int)

The number of targets for which the status is Delivery Timed Out.

command_timeoutSeconds :: Lens' Command (Maybe Natural)

The TimeoutSeconds value specified for a command.

command_outputS3KeyPrefix :: Lens' Command (Maybe Text)

The S3 directory path inside the bucket where the responses to the command executions should be stored. This was requested when issuing the command.

command_cloudWatchOutputConfig :: Lens' Command (Maybe CloudWatchOutputConfig)

Amazon CloudWatch Logs information where you want Amazon Web Services Systems Manager to send the command output.

command_documentVersion :: Lens' Command (Maybe Text)

The Systems Manager document (SSM document) version.

command_targetCount :: Lens' Command (Maybe Int)

The number of targets for the command.

command_parameters :: Lens' Command (Maybe (HashMap Text [Text]))

The parameter values to be inserted in the document when running the command.