Network.AWS.SSM.Types.CommandFilter

Description

Documentation

data CommandFilter

Describes a command filter.

An instance ID can't be specified when a command status is Pending because the command hasn't run on the instance yet.

See: newCommandFilter smart constructor.

Constructors

CommandFilter'

Fields

key :: CommandFilterKey

The name of the filter.

value :: Text

The filter value. Valid values for each filter key are as follows:

InvokedAfter: Specify a timestamp to limit your results. For example, specify 2021-07-07T00:00:00Z to see a list of command executions occurring July 7, 2021, and later.

InvokedBefore: Specify a timestamp to limit your results. For example, specify 2021-07-07T00:00:00Z to see a list of command executions from before July 7, 2021.

Status: Specify a valid command status to see a list of all command executions with that status. The status choices depend on the API you call.

The status values you can specify for ListCommands are:

Pending

InProgress

Success

Cancelled

Failed

TimedOut (this includes both Delivery and Execution time outs)

AccessDenied

DeliveryTimedOut

ExecutionTimedOut

Incomplete

NoInstancesInTag

LimitExceeded

The status values you can specify for ListCommandInvocations are:

Pending

InProgress

Delayed

Success

Cancelled

Failed

TimedOut (this includes both Delivery and Execution time outs)

AccessDenied

DeliveryTimedOut

ExecutionTimedOut

Undeliverable

InvalidPlatform

Terminated

DocumentName: Specify name of the Amazon Web Services Systems Manager document (SSM document) for which you want to see command execution results. For example, specify AWS-RunPatchBaseline to see command executions that used this SSM document to perform security patching operations on instances.

ExecutionStage: Specify one of the following values:

Executing: Returns a list of command executions that are currently still running.

Complete: Returns a list of command executions that have already completed.

newCommandFilter

Arguments

:: CommandFilterKey
$sel:key:CommandFilter'

-> Text
$sel:value:CommandFilter'

-> CommandFilter

Create a value of CommandFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:CommandFilter', commandFilter_key - The name of the filter.

$sel:value:CommandFilter', commandFilter_value - The filter value. Valid values for each filter key are as follows:

InvokedAfter: Specify a timestamp to limit your results. For example, specify 2021-07-07T00:00:00Z to see a list of command executions occurring July 7, 2021, and later.

InvokedBefore: Specify a timestamp to limit your results. For example, specify 2021-07-07T00:00:00Z to see a list of command executions from before July 7, 2021.

Status: Specify a valid command status to see a list of all command executions with that status. The status choices depend on the API you call.

The status values you can specify for ListCommands are:

Pending

InProgress

Success

Cancelled

Failed

TimedOut (this includes both Delivery and Execution time outs)

AccessDenied

DeliveryTimedOut

ExecutionTimedOut

Incomplete

NoInstancesInTag

LimitExceeded

The status values you can specify for ListCommandInvocations are:

Pending

InProgress

Delayed

Success

Cancelled

Failed

TimedOut (this includes both Delivery and Execution time outs)

AccessDenied

DeliveryTimedOut

ExecutionTimedOut

Undeliverable

InvalidPlatform

Terminated

DocumentName: Specify name of the Amazon Web Services Systems Manager document (SSM document) for which you want to see command execution results. For example, specify AWS-RunPatchBaseline to see command executions that used this SSM document to perform security patching operations on instances.

ExecutionStage: Specify one of the following values:

Executing: Returns a list of command executions that are currently still running.

Complete: Returns a list of command executions that have already completed.

commandFilter_key :: Lens' CommandFilter CommandFilterKey

The name of the filter.

commandFilter_value :: Lens' CommandFilter Text

The filter value. Valid values for each filter key are as follows:

InvokedAfter: Specify a timestamp to limit your results. For example, specify 2021-07-07T00:00:00Z to see a list of command executions occurring July 7, 2021, and later.

InvokedBefore: Specify a timestamp to limit your results. For example, specify 2021-07-07T00:00:00Z to see a list of command executions from before July 7, 2021.

Status: Specify a valid command status to see a list of all command executions with that status. The status choices depend on the API you call.

The status values you can specify for ListCommands are:

Pending

InProgress

Success

Cancelled

Failed

TimedOut (this includes both Delivery and Execution time outs)

AccessDenied

DeliveryTimedOut

ExecutionTimedOut

Incomplete

NoInstancesInTag

LimitExceeded

The status values you can specify for ListCommandInvocations are:

Pending

InProgress

Delayed

Success

Cancelled

Failed

TimedOut (this includes both Delivery and Execution time outs)

AccessDenied

DeliveryTimedOut

ExecutionTimedOut

Undeliverable

InvalidPlatform

Terminated

DocumentName: Specify name of the Amazon Web Services Systems Manager document (SSM document) for which you want to see command execution results. For example, specify AWS-RunPatchBaseline to see command executions that used this SSM document to perform security patching operations on instances.

ExecutionStage: Specify one of the following values:

Executing: Returns a list of command executions that are currently still running.

Complete: Returns a list of command executions that have already completed.