Network.AWS.SSM.Types.CommandInvocation

Description

Documentation

data CommandInvocation

An invocation is copy of a command sent to a specific instance. A command can apply to one or more instances. A command invocation applies to one instance. For example, if a user runs SendCommand against three instances, then a command invocation is created for each requested instance ID. A command invocation returns status and detail information about a command you ran.

See: newCommandInvocation smart constructor.

Constructors

CommandInvocation'

Fields

notificationConfig :: Maybe NotificationConfig

Configurations for sending notifications about command status changes on a per instance basis.

standardOutputUrl :: Maybe Text

The URL to the plugin's StdOut file in Amazon Simple Storage Service (Amazon S3), if the S3 bucket was defined for the parent command. For an invocation, StandardOutputUrl is populated if there is just one plugin defined for the command, and the S3 bucket was defined for the command.

status :: Maybe CommandInvocationStatus

Whether or not the invocation succeeded, failed, or is pending.

instanceId :: Maybe Text

The instance ID in which this invocation was requested.

requestedDateTime :: Maybe POSIX

The time and date the request was sent to this instance.

serviceRole :: Maybe Text

The Identity and Access Management (IAM) service role that Run Command, a capability of Amazon Web Services Systems Manager, uses to act on your behalf when sending notifications about command status changes on a per instance basis.

statusDetails :: Maybe Text

A detailed status of the command execution for each invocation (each instance targeted by the command). StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Understanding command statuses in the Amazon Web Services Systems Manager User Guide. StatusDetails can be one of the following values:

Pending: The command hasn't been sent to the instance.

In Progress: The command has been sent to the instance but hasn't reached a terminal state.

Success: The execution of the command or plugin was successfully completed. This is a terminal state.

Delivery Timed Out: The command wasn't delivered to the instance before the delivery timeout expired. Delivery timeouts don't count against the parent command's MaxErrors limit, but they do contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Execution Timed Out: Command execution started on the instance, but the execution wasn't complete before the execution timeout expired. Execution timeouts count against the MaxErrors limit of the parent command. This is a terminal state.

Failed: The command wasn't successful on the instance. For a plugin, this indicates that the result code wasn't zero. For a command invocation, this indicates that the result code for one or more plugins wasn't zero. Invocation failures count against the MaxErrors limit of the parent command. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Undeliverable: The command can't be delivered to the instance. The instance might not exist or might not be responding. Undeliverable invocations don't count against the parent command's MaxErrors limit and don't contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Terminated: The parent command exceeded its MaxErrors limit and subsequent command invocations were canceled by the system. This is a terminal state.

instanceName :: Maybe Text

The fully qualified host name of the managed instance.

comment :: Maybe Text

User-specified information about the command, such as a brief description of what the command should do.

standardErrorUrl :: Maybe Text

The URL to the plugin's StdErr file in Amazon Simple Storage Service (Amazon S3), if the S3 bucket was defined for the parent command. For an invocation, StandardErrorUrl is populated if there is just one plugin defined for the command, and the S3 bucket was defined for the command.

documentName :: Maybe Text

The document name that was requested for execution.

commandId :: Maybe Text

The command against which this invocation was requested.

traceOutput :: Maybe Text

Gets the trace output sent by the agent.

commandPlugins :: Maybe [CommandPlugin]

Plugins processed by the command.

cloudWatchOutputConfig :: Maybe CloudWatchOutputConfig

Amazon CloudWatch Logs information where you want Amazon Web Services Systems Manager to send the command output.

documentVersion :: Maybe Text

The Systems Manager document (SSM document) version.

newCommandInvocation :: CommandInvocation

Create a value of CommandInvocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notificationConfig:CommandInvocation', commandInvocation_notificationConfig - Configurations for sending notifications about command status changes on a per instance basis.

$sel:standardOutputUrl:CommandInvocation', commandInvocation_standardOutputUrl - The URL to the plugin's StdOut file in Amazon Simple Storage Service (Amazon S3), if the S3 bucket was defined for the parent command. For an invocation, StandardOutputUrl is populated if there is just one plugin defined for the command, and the S3 bucket was defined for the command.

$sel:status:CommandInvocation', commandInvocation_status - Whether or not the invocation succeeded, failed, or is pending.

$sel:instanceId:CommandInvocation', commandInvocation_instanceId - The instance ID in which this invocation was requested.

$sel:requestedDateTime:CommandInvocation', commandInvocation_requestedDateTime - The time and date the request was sent to this instance.

$sel:serviceRole:CommandInvocation', commandInvocation_serviceRole - The Identity and Access Management (IAM) service role that Run Command, a capability of Amazon Web Services Systems Manager, uses to act on your behalf when sending notifications about command status changes on a per instance basis.

$sel:statusDetails:CommandInvocation', commandInvocation_statusDetails - A detailed status of the command execution for each invocation (each instance targeted by the command). StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Understanding command statuses in the Amazon Web Services Systems Manager User Guide. StatusDetails can be one of the following values:

Pending: The command hasn't been sent to the instance.

In Progress: The command has been sent to the instance but hasn't reached a terminal state.

Success: The execution of the command or plugin was successfully completed. This is a terminal state.

Delivery Timed Out: The command wasn't delivered to the instance before the delivery timeout expired. Delivery timeouts don't count against the parent command's MaxErrors limit, but they do contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Execution Timed Out: Command execution started on the instance, but the execution wasn't complete before the execution timeout expired. Execution timeouts count against the MaxErrors limit of the parent command. This is a terminal state.

Failed: The command wasn't successful on the instance. For a plugin, this indicates that the result code wasn't zero. For a command invocation, this indicates that the result code for one or more plugins wasn't zero. Invocation failures count against the MaxErrors limit of the parent command. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Undeliverable: The command can't be delivered to the instance. The instance might not exist or might not be responding. Undeliverable invocations don't count against the parent command's MaxErrors limit and don't contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Terminated: The parent command exceeded its MaxErrors limit and subsequent command invocations were canceled by the system. This is a terminal state.

$sel:instanceName:CommandInvocation', commandInvocation_instanceName - The fully qualified host name of the managed instance.

$sel:comment:CommandInvocation', commandInvocation_comment - User-specified information about the command, such as a brief description of what the command should do.

$sel:standardErrorUrl:CommandInvocation', commandInvocation_standardErrorUrl - The URL to the plugin's StdErr file in Amazon Simple Storage Service (Amazon S3), if the S3 bucket was defined for the parent command. For an invocation, StandardErrorUrl is populated if there is just one plugin defined for the command, and the S3 bucket was defined for the command.

$sel:documentName:CommandInvocation', commandInvocation_documentName - The document name that was requested for execution.

$sel:commandId:CommandInvocation', commandInvocation_commandId - The command against which this invocation was requested.

$sel:traceOutput:CommandInvocation', commandInvocation_traceOutput - Gets the trace output sent by the agent.

$sel:commandPlugins:CommandInvocation', commandInvocation_commandPlugins - Plugins processed by the command.

$sel:cloudWatchOutputConfig:CommandInvocation', commandInvocation_cloudWatchOutputConfig - Amazon CloudWatch Logs information where you want Amazon Web Services Systems Manager to send the command output.

$sel:documentVersion:CommandInvocation', commandInvocation_documentVersion - The Systems Manager document (SSM document) version.

commandInvocation_notificationConfig :: Lens' CommandInvocation (Maybe NotificationConfig)

Configurations for sending notifications about command status changes on a per instance basis.

commandInvocation_standardOutputUrl :: Lens' CommandInvocation (Maybe Text)

The URL to the plugin's StdOut file in Amazon Simple Storage Service (Amazon S3), if the S3 bucket was defined for the parent command. For an invocation, StandardOutputUrl is populated if there is just one plugin defined for the command, and the S3 bucket was defined for the command.

commandInvocation_status :: Lens' CommandInvocation (Maybe CommandInvocationStatus)

Whether or not the invocation succeeded, failed, or is pending.

commandInvocation_instanceId :: Lens' CommandInvocation (Maybe Text)

The instance ID in which this invocation was requested.

commandInvocation_requestedDateTime :: Lens' CommandInvocation (Maybe UTCTime)

The time and date the request was sent to this instance.

commandInvocation_serviceRole :: Lens' CommandInvocation (Maybe Text)

The Identity and Access Management (IAM) service role that Run Command, a capability of Amazon Web Services Systems Manager, uses to act on your behalf when sending notifications about command status changes on a per instance basis.

commandInvocation_statusDetails :: Lens' CommandInvocation (Maybe Text)

A detailed status of the command execution for each invocation (each instance targeted by the command). StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Understanding command statuses in the Amazon Web Services Systems Manager User Guide. StatusDetails can be one of the following values:

Pending: The command hasn't been sent to the instance.

In Progress: The command has been sent to the instance but hasn't reached a terminal state.

Success: The execution of the command or plugin was successfully completed. This is a terminal state.

Delivery Timed Out: The command wasn't delivered to the instance before the delivery timeout expired. Delivery timeouts don't count against the parent command's MaxErrors limit, but they do contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Execution Timed Out: Command execution started on the instance, but the execution wasn't complete before the execution timeout expired. Execution timeouts count against the MaxErrors limit of the parent command. This is a terminal state.

Failed: The command wasn't successful on the instance. For a plugin, this indicates that the result code wasn't zero. For a command invocation, this indicates that the result code for one or more plugins wasn't zero. Invocation failures count against the MaxErrors limit of the parent command. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Undeliverable: The command can't be delivered to the instance. The instance might not exist or might not be responding. Undeliverable invocations don't count against the parent command's MaxErrors limit and don't contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Terminated: The parent command exceeded its MaxErrors limit and subsequent command invocations were canceled by the system. This is a terminal state.

commandInvocation_instanceName :: Lens' CommandInvocation (Maybe Text)

The fully qualified host name of the managed instance.

commandInvocation_comment :: Lens' CommandInvocation (Maybe Text)

User-specified information about the command, such as a brief description of what the command should do.

commandInvocation_standardErrorUrl :: Lens' CommandInvocation (Maybe Text)

The URL to the plugin's StdErr file in Amazon Simple Storage Service (Amazon S3), if the S3 bucket was defined for the parent command. For an invocation, StandardErrorUrl is populated if there is just one plugin defined for the command, and the S3 bucket was defined for the command.

commandInvocation_documentName :: Lens' CommandInvocation (Maybe Text)

The document name that was requested for execution.

commandInvocation_commandId :: Lens' CommandInvocation (Maybe Text)

The command against which this invocation was requested.

commandInvocation_traceOutput :: Lens' CommandInvocation (Maybe Text)

Gets the trace output sent by the agent.

commandInvocation_commandPlugins :: Lens' CommandInvocation (Maybe [CommandPlugin])

Plugins processed by the command.

commandInvocation_cloudWatchOutputConfig :: Lens' CommandInvocation (Maybe CloudWatchOutputConfig)

Amazon CloudWatch Logs information where you want Amazon Web Services Systems Manager to send the command output.

commandInvocation_documentVersion :: Lens' CommandInvocation (Maybe Text)

The Systems Manager document (SSM document) version.