Network.AWS.SSM.Types.CommandPlugin

Description

Documentation

data CommandPlugin

Describes plugin details.

See: newCommandPlugin smart constructor.

Constructors

CommandPlugin'

Fields

status :: Maybe CommandPluginStatus

The status of this plugin. You can run a document with multiple plugins.

responseStartDateTime :: Maybe POSIX

The time the plugin started running.

outputS3KeyPrefix :: Maybe Text

The S3 directory path inside the bucket where the responses to the command executions should be stored. This was requested when issuing the command. For example, in the following response:

doc-example-bucket/ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix/i-02573cafcfEXAMPLE/awsrunShellScript

doc-example-bucket is the name of the S3 bucket;

ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix is the name of the S3 prefix;

i-02573cafcfEXAMPLE is the instance ID;

awsrunShellScript is the name of the plugin.

standardErrorUrl :: Maybe Text

The URL for the complete text written by the plugin to stderr. If execution isn't yet complete, then this string is empty.

responseCode :: Maybe Int

A numeric response code generated after running the plugin.

statusDetails :: Maybe Text

A detailed status of the plugin execution. StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Understanding command statuses in the Amazon Web Services Systems Manager User Guide. StatusDetails can be one of the following values:

Pending: The command hasn't been sent to the instance.

In Progress: The command has been sent to the instance but hasn't reached a terminal state.

Success: The execution of the command or plugin was successfully completed. This is a terminal state.

Delivery Timed Out: The command wasn't delivered to the instance before the delivery timeout expired. Delivery timeouts don't count against the parent command's MaxErrors limit, but they do contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Execution Timed Out: Command execution started on the instance, but the execution wasn't complete before the execution timeout expired. Execution timeouts count against the MaxErrors limit of the parent command. This is a terminal state.

Failed: The command wasn't successful on the instance. For a plugin, this indicates that the result code wasn't zero. For a command invocation, this indicates that the result code for one or more plugins wasn't zero. Invocation failures count against the MaxErrors limit of the parent command. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Undeliverable: The command can't be delivered to the instance. The instance might not exist, or it might not be responding. Undeliverable invocations don't count against the parent command's MaxErrors limit, and they don't contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Terminated: The parent command exceeded its MaxErrors limit and subsequent command invocations were canceled by the system. This is a terminal state.

output :: Maybe Text

Output of the plugin execution.

standardOutputUrl :: Maybe Text

The URL for the complete text written by the plugin to stdout in Amazon S3. If the S3 bucket for the command wasn't specified, then this string is empty.

name :: Maybe Text

The name of the plugin. Must be one of the following: aws:updateAgent, aws:domainjoin, aws:applications, aws:runPowerShellScript, aws:psmodule, aws:cloudWatch, aws:runShellScript, or aws:updateSSMAgent.

outputS3Region :: Maybe Text

(Deprecated) You can no longer specify this parameter. The system ignores it. Instead, Amazon Web Services Systems Manager automatically determines the S3 bucket region.

outputS3BucketName :: Maybe Text

The S3 bucket where the responses to the command executions should be stored. This was requested when issuing the command. For example, in the following response:

doc-example-bucket/ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix/i-02573cafcfEXAMPLE/awsrunShellScript

doc-example-bucket is the name of the S3 bucket;

ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix is the name of the S3 prefix;

i-02573cafcfEXAMPLE is the instance ID;

awsrunShellScript is the name of the plugin.

responseFinishDateTime :: Maybe POSIX

The time the plugin stopped running. Could stop prematurely if, for example, a cancel command was sent.

newCommandPlugin :: CommandPlugin

Create a value of CommandPlugin with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CommandPlugin', commandPlugin_status - The status of this plugin. You can run a document with multiple plugins.

$sel:responseStartDateTime:CommandPlugin', commandPlugin_responseStartDateTime - The time the plugin started running.

$sel:outputS3KeyPrefix:CommandPlugin', commandPlugin_outputS3KeyPrefix - The S3 directory path inside the bucket where the responses to the command executions should be stored. This was requested when issuing the command. For example, in the following response:

doc-example-bucket/ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix/i-02573cafcfEXAMPLE/awsrunShellScript

doc-example-bucket is the name of the S3 bucket;

ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix is the name of the S3 prefix;

i-02573cafcfEXAMPLE is the instance ID;

awsrunShellScript is the name of the plugin.

$sel:standardErrorUrl:CommandPlugin', commandPlugin_standardErrorUrl - The URL for the complete text written by the plugin to stderr. If execution isn't yet complete, then this string is empty.

$sel:responseCode:CommandPlugin', commandPlugin_responseCode - A numeric response code generated after running the plugin.

$sel:statusDetails:CommandPlugin', commandPlugin_statusDetails - A detailed status of the plugin execution. StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Understanding command statuses in the Amazon Web Services Systems Manager User Guide. StatusDetails can be one of the following values:

Pending: The command hasn't been sent to the instance.

In Progress: The command has been sent to the instance but hasn't reached a terminal state.

Success: The execution of the command or plugin was successfully completed. This is a terminal state.

Delivery Timed Out: The command wasn't delivered to the instance before the delivery timeout expired. Delivery timeouts don't count against the parent command's MaxErrors limit, but they do contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Execution Timed Out: Command execution started on the instance, but the execution wasn't complete before the execution timeout expired. Execution timeouts count against the MaxErrors limit of the parent command. This is a terminal state.

Failed: The command wasn't successful on the instance. For a plugin, this indicates that the result code wasn't zero. For a command invocation, this indicates that the result code for one or more plugins wasn't zero. Invocation failures count against the MaxErrors limit of the parent command. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Undeliverable: The command can't be delivered to the instance. The instance might not exist, or it might not be responding. Undeliverable invocations don't count against the parent command's MaxErrors limit, and they don't contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Terminated: The parent command exceeded its MaxErrors limit and subsequent command invocations were canceled by the system. This is a terminal state.

$sel:output:CommandPlugin', commandPlugin_output - Output of the plugin execution.

$sel:standardOutputUrl:CommandPlugin', commandPlugin_standardOutputUrl - The URL for the complete text written by the plugin to stdout in Amazon S3. If the S3 bucket for the command wasn't specified, then this string is empty.

$sel:name:CommandPlugin', commandPlugin_name - The name of the plugin. Must be one of the following: aws:updateAgent, aws:domainjoin, aws:applications, aws:runPowerShellScript, aws:psmodule, aws:cloudWatch, aws:runShellScript, or aws:updateSSMAgent.

$sel:outputS3Region:CommandPlugin', commandPlugin_outputS3Region - (Deprecated) You can no longer specify this parameter. The system ignores it. Instead, Amazon Web Services Systems Manager automatically determines the S3 bucket region.

$sel:outputS3BucketName:CommandPlugin', commandPlugin_outputS3BucketName - The S3 bucket where the responses to the command executions should be stored. This was requested when issuing the command. For example, in the following response:

doc-example-bucket/ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix/i-02573cafcfEXAMPLE/awsrunShellScript

doc-example-bucket is the name of the S3 bucket;

ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix is the name of the S3 prefix;

i-02573cafcfEXAMPLE is the instance ID;

awsrunShellScript is the name of the plugin.

$sel:responseFinishDateTime:CommandPlugin', commandPlugin_responseFinishDateTime - The time the plugin stopped running. Could stop prematurely if, for example, a cancel command was sent.

commandPlugin_status :: Lens' CommandPlugin (Maybe CommandPluginStatus)

The status of this plugin. You can run a document with multiple plugins.

commandPlugin_responseStartDateTime :: Lens' CommandPlugin (Maybe UTCTime)

The time the plugin started running.

commandPlugin_outputS3KeyPrefix :: Lens' CommandPlugin (Maybe Text)

The S3 directory path inside the bucket where the responses to the command executions should be stored. This was requested when issuing the command. For example, in the following response:

doc-example-bucket/ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix/i-02573cafcfEXAMPLE/awsrunShellScript

doc-example-bucket is the name of the S3 bucket;

ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix is the name of the S3 prefix;

i-02573cafcfEXAMPLE is the instance ID;

awsrunShellScript is the name of the plugin.

commandPlugin_standardErrorUrl :: Lens' CommandPlugin (Maybe Text)

The URL for the complete text written by the plugin to stderr. If execution isn't yet complete, then this string is empty.

commandPlugin_responseCode :: Lens' CommandPlugin (Maybe Int)

A numeric response code generated after running the plugin.

commandPlugin_statusDetails :: Lens' CommandPlugin (Maybe Text)

A detailed status of the plugin execution. StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Understanding command statuses in the Amazon Web Services Systems Manager User Guide. StatusDetails can be one of the following values:

Pending: The command hasn't been sent to the instance.

In Progress: The command has been sent to the instance but hasn't reached a terminal state.

Success: The execution of the command or plugin was successfully completed. This is a terminal state.

Delivery Timed Out: The command wasn't delivered to the instance before the delivery timeout expired. Delivery timeouts don't count against the parent command's MaxErrors limit, but they do contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Execution Timed Out: Command execution started on the instance, but the execution wasn't complete before the execution timeout expired. Execution timeouts count against the MaxErrors limit of the parent command. This is a terminal state.

Failed: The command wasn't successful on the instance. For a plugin, this indicates that the result code wasn't zero. For a command invocation, this indicates that the result code for one or more plugins wasn't zero. Invocation failures count against the MaxErrors limit of the parent command. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Undeliverable: The command can't be delivered to the instance. The instance might not exist, or it might not be responding. Undeliverable invocations don't count against the parent command's MaxErrors limit, and they don't contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Terminated: The parent command exceeded its MaxErrors limit and subsequent command invocations were canceled by the system. This is a terminal state.

commandPlugin_output :: Lens' CommandPlugin (Maybe Text)

Output of the plugin execution.

commandPlugin_standardOutputUrl :: Lens' CommandPlugin (Maybe Text)

The URL for the complete text written by the plugin to stdout in Amazon S3. If the S3 bucket for the command wasn't specified, then this string is empty.

commandPlugin_name :: Lens' CommandPlugin (Maybe Text)

The name of the plugin. Must be one of the following: aws:updateAgent, aws:domainjoin, aws:applications, aws:runPowerShellScript, aws:psmodule, aws:cloudWatch, aws:runShellScript, or aws:updateSSMAgent.

commandPlugin_outputS3Region :: Lens' CommandPlugin (Maybe Text)

(Deprecated) You can no longer specify this parameter. The system ignores it. Instead, Amazon Web Services Systems Manager automatically determines the S3 bucket region.

commandPlugin_outputS3BucketName :: Lens' CommandPlugin (Maybe Text)

The S3 bucket where the responses to the command executions should be stored. This was requested when issuing the command. For example, in the following response:

doc-example-bucket/ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix/i-02573cafcfEXAMPLE/awsrunShellScript

doc-example-bucket is the name of the S3 bucket;

ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix is the name of the S3 prefix;

i-02573cafcfEXAMPLE is the instance ID;

awsrunShellScript is the name of the plugin.

commandPlugin_responseFinishDateTime :: Lens' CommandPlugin (Maybe UTCTime)

The time the plugin stopped running. Could stop prematurely if, for example, a cancel command was sent.