Network.AWS.SSM.Types.ComplianceItem

Description

Documentation

data ComplianceItem

Information about the compliance as defined by the resource type. For example, for a patch resource type, Items includes information about the PatchSeverity, Classification, and so on.

See: newComplianceItem smart constructor.

Constructors

ComplianceItem'

Fields

resourceId :: Maybe Text

An ID for the resource. For a managed instance, this is the instance ID.

status :: Maybe ComplianceStatus

The status of the compliance item. An item is either COMPLIANT, NON_COMPLIANT, or an empty string (for Windows patches that aren't applicable).

severity :: Maybe ComplianceSeverity

The severity of the compliance status. Severity can be one of the following: Critical, High, Medium, Low, Informational, Unspecified.

title :: Maybe Text

A title for the compliance item. For example, if the compliance item is a Windows patch, the title could be the title of the KB article for the patch; for example: Security Update for Active Directory Federation Services.

complianceType :: Maybe Text

The compliance type. For example, Association (for a State Manager association), Patch, or Custom:string are all valid compliance types.

id :: Maybe Text

An ID for the compliance item. For example, if the compliance item is a Windows patch, the ID could be the number of the KB article; for example: KB4010320.

resourceType :: Maybe Text

The type of resource. ManagedInstance is currently the only supported resource type.

details :: Maybe (HashMap Text Text)

A "Key": "Value" tag combination for the compliance item.

executionSummary :: Maybe ComplianceExecutionSummary

A summary for the compliance item. The summary includes an execution ID, the execution type (for example, command), and the execution time.

newComplianceItem :: ComplianceItem

Create a value of ComplianceItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:ComplianceItem', complianceItem_resourceId - An ID for the resource. For a managed instance, this is the instance ID.

$sel:status:ComplianceItem', complianceItem_status - The status of the compliance item. An item is either COMPLIANT, NON_COMPLIANT, or an empty string (for Windows patches that aren't applicable).

$sel:severity:ComplianceItem', complianceItem_severity - The severity of the compliance status. Severity can be one of the following: Critical, High, Medium, Low, Informational, Unspecified.

$sel:title:ComplianceItem', complianceItem_title - A title for the compliance item. For example, if the compliance item is a Windows patch, the title could be the title of the KB article for the patch; for example: Security Update for Active Directory Federation Services.

$sel:complianceType:ComplianceItem', complianceItem_complianceType - The compliance type. For example, Association (for a State Manager association), Patch, or Custom:string are all valid compliance types.

$sel:id:ComplianceItem', complianceItem_id - An ID for the compliance item. For example, if the compliance item is a Windows patch, the ID could be the number of the KB article; for example: KB4010320.

$sel:resourceType:ComplianceItem', complianceItem_resourceType - The type of resource. ManagedInstance is currently the only supported resource type.

$sel:details:ComplianceItem', complianceItem_details - A "Key": "Value" tag combination for the compliance item.

$sel:executionSummary:ComplianceItem', complianceItem_executionSummary - A summary for the compliance item. The summary includes an execution ID, the execution type (for example, command), and the execution time.

complianceItem_resourceId :: Lens' ComplianceItem (Maybe Text)

An ID for the resource. For a managed instance, this is the instance ID.

complianceItem_status :: Lens' ComplianceItem (Maybe ComplianceStatus)

The status of the compliance item. An item is either COMPLIANT, NON_COMPLIANT, or an empty string (for Windows patches that aren't applicable).

complianceItem_severity :: Lens' ComplianceItem (Maybe ComplianceSeverity)

The severity of the compliance status. Severity can be one of the following: Critical, High, Medium, Low, Informational, Unspecified.

complianceItem_title :: Lens' ComplianceItem (Maybe Text)

A title for the compliance item. For example, if the compliance item is a Windows patch, the title could be the title of the KB article for the patch; for example: Security Update for Active Directory Federation Services.

complianceItem_complianceType :: Lens' ComplianceItem (Maybe Text)

The compliance type. For example, Association (for a State Manager association), Patch, or Custom:string are all valid compliance types.

complianceItem_id :: Lens' ComplianceItem (Maybe Text)

An ID for the compliance item. For example, if the compliance item is a Windows patch, the ID could be the number of the KB article; for example: KB4010320.

complianceItem_resourceType :: Lens' ComplianceItem (Maybe Text)

The type of resource. ManagedInstance is currently the only supported resource type.

complianceItem_details :: Lens' ComplianceItem (Maybe (HashMap Text Text))

A "Key": "Value" tag combination for the compliance item.

complianceItem_executionSummary :: Lens' ComplianceItem (Maybe ComplianceExecutionSummary)

A summary for the compliance item. The summary includes an execution ID, the execution type (for example, command), and the execution time.