Network.AWS.SSM.Types.ComplianceItemEntry

Description

Documentation

data ComplianceItemEntry

Information about a compliance item.

See: newComplianceItemEntry smart constructor.

Constructors

ComplianceItemEntry'

Fields

title :: Maybe Text

The title of the compliance item. For example, if the compliance item is a Windows patch, the title could be the title of the KB article for the patch; for example: Security Update for Active Directory Federation Services.

id :: Maybe Text

The compliance item ID. For example, if the compliance item is a Windows patch, the ID could be the number of the KB article.

details :: Maybe (HashMap Text Text)

A "Key": "Value" tag combination for the compliance item.

severity :: ComplianceSeverity

The severity of the compliance status. Severity can be one of the following: Critical, High, Medium, Low, Informational, Unspecified.

status :: ComplianceStatus

The status of the compliance item. An item is either COMPLIANT or NON_COMPLIANT.

newComplianceItemEntry

Arguments

:: ComplianceSeverity
$sel:severity:ComplianceItemEntry'

-> ComplianceStatus
$sel:status:ComplianceItemEntry'

-> ComplianceItemEntry

Create a value of ComplianceItemEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:title:ComplianceItemEntry', complianceItemEntry_title - The title of the compliance item. For example, if the compliance item is a Windows patch, the title could be the title of the KB article for the patch; for example: Security Update for Active Directory Federation Services.

$sel:id:ComplianceItemEntry', complianceItemEntry_id - The compliance item ID. For example, if the compliance item is a Windows patch, the ID could be the number of the KB article.

$sel:details:ComplianceItemEntry', complianceItemEntry_details - A "Key": "Value" tag combination for the compliance item.

$sel:severity:ComplianceItemEntry', complianceItemEntry_severity - The severity of the compliance status. Severity can be one of the following: Critical, High, Medium, Low, Informational, Unspecified.

$sel:status:ComplianceItemEntry', complianceItemEntry_status - The status of the compliance item. An item is either COMPLIANT or NON_COMPLIANT.

complianceItemEntry_title :: Lens' ComplianceItemEntry (Maybe Text)

The title of the compliance item. For example, if the compliance item is a Windows patch, the title could be the title of the KB article for the patch; for example: Security Update for Active Directory Federation Services.

complianceItemEntry_id :: Lens' ComplianceItemEntry (Maybe Text)

The compliance item ID. For example, if the compliance item is a Windows patch, the ID could be the number of the KB article.

complianceItemEntry_details :: Lens' ComplianceItemEntry (Maybe (HashMap Text Text))

A "Key": "Value" tag combination for the compliance item.

complianceItemEntry_severity :: Lens' ComplianceItemEntry ComplianceSeverity

The severity of the compliance status. Severity can be one of the following: Critical, High, Medium, Low, Informational, Unspecified.

complianceItemEntry_status :: Lens' ComplianceItemEntry ComplianceStatus

The status of the compliance item. An item is either COMPLIANT or NON_COMPLIANT.