Network.AWS.SSM.Types.DocumentDescription

Description

Documentation

data DocumentDescription

Describes a Amazon Web Services Systems Manager document (SSM document).

See: newDocumentDescription smart constructor.

Constructors

DocumentDescription'

Fields

documentType :: Maybe DocumentType

The type of document.

platformTypes :: Maybe [PlatformType]

The list of OS platforms compatible with this SSM document.

createdDate :: Maybe POSIX

The date when the document was created.

status :: Maybe DocumentStatus

The status of the SSM document.

approvedVersion :: Maybe Text

The version of the document currently approved for use in the organization.

defaultVersion :: Maybe Text

The default version.

latestVersion :: Maybe Text

The latest version of the document.

targetType :: Maybe Text

The target type which defines the kinds of resources the document can run on. For example, /AWS::EC2::Instance. For a list of valid resource types, see Amazon Web Services resource and property types reference in the CloudFormation User Guide.

requires :: Maybe (NonEmpty DocumentRequires)

A list of SSM documents required by a document. For example, an ApplicationConfiguration document requires an ApplicationConfigurationSchema document.

sha1 :: Maybe Text

The SHA1 hash of the document, which you can use for verification.

statusInformation :: Maybe Text

A message returned by Amazon Web Services Systems Manager that explains the Status value. For example, a Failed status might be explained by the StatusInformation message, "The specified S3 bucket doesn't exist. Verify that the URL of the S3 bucket is correct."

author :: Maybe Text

The user in your organization who created the document.

versionName :: Maybe Text

The version of the artifact associated with the document.

hash :: Maybe Text

The Sha256 or Sha1 hash created by the system when the document was created.

Sha1 hashes have been deprecated.

name :: Maybe Text

The name of the SSM document.

pendingReviewVersion :: Maybe Text

The version of the document that is currently under review.

documentFormat :: Maybe DocumentFormat

The document format, either JSON or YAML.

tags :: Maybe [Tag]

The tags, or metadata, that have been applied to the document.

owner :: Maybe Text

The Amazon Web Services user account that created the document.

attachmentsInformation :: Maybe [AttachmentInformation]

Details about the document attachments, including names, locations, sizes, and so on.

reviewInformation :: Maybe (NonEmpty ReviewInformation)

Details about the review of a document.

description :: Maybe Text

A description of the document.

reviewStatus :: Maybe ReviewStatus

The current status of the review.

schemaVersion :: Maybe Text

The schema version.

documentVersion :: Maybe Text

The document version.

displayName :: Maybe Text

The friendly name of the SSM document. This value can differ for each version of the document. If you want to update this value, see UpdateDocument.

hashType :: Maybe DocumentHashType

The hash type of the document. Valid values include Sha256 or Sha1.

Sha1 hashes have been deprecated.

parameters :: Maybe [DocumentParameter]

A description of the parameters for a document.

newDocumentDescription :: DocumentDescription

Create a value of DocumentDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:documentType:DocumentDescription', documentDescription_documentType - The type of document.

$sel:platformTypes:DocumentDescription', documentDescription_platformTypes - The list of OS platforms compatible with this SSM document.

$sel:createdDate:DocumentDescription', documentDescription_createdDate - The date when the document was created.

$sel:status:DocumentDescription', documentDescription_status - The status of the SSM document.

$sel:approvedVersion:DocumentDescription', documentDescription_approvedVersion - The version of the document currently approved for use in the organization.

$sel:defaultVersion:DocumentDescription', documentDescription_defaultVersion - The default version.

$sel:latestVersion:DocumentDescription', documentDescription_latestVersion - The latest version of the document.

$sel:targetType:DocumentDescription', documentDescription_targetType - The target type which defines the kinds of resources the document can run on. For example, /AWS::EC2::Instance. For a list of valid resource types, see Amazon Web Services resource and property types reference in the CloudFormation User Guide.

$sel:requires:DocumentDescription', documentDescription_requires - A list of SSM documents required by a document. For example, an ApplicationConfiguration document requires an ApplicationConfigurationSchema document.

$sel:sha1:DocumentDescription', documentDescription_sha1 - The SHA1 hash of the document, which you can use for verification.

$sel:statusInformation:DocumentDescription', documentDescription_statusInformation - A message returned by Amazon Web Services Systems Manager that explains the Status value. For example, a Failed status might be explained by the StatusInformation message, "The specified S3 bucket doesn't exist. Verify that the URL of the S3 bucket is correct."

$sel:author:DocumentDescription', documentDescription_author - The user in your organization who created the document.

$sel:versionName:DocumentDescription', documentDescription_versionName - The version of the artifact associated with the document.

$sel:hash:DocumentDescription', documentDescription_hash - The Sha256 or Sha1 hash created by the system when the document was created.

Sha1 hashes have been deprecated.

$sel:name:DocumentDescription', documentDescription_name - The name of the SSM document.

$sel:pendingReviewVersion:DocumentDescription', documentDescription_pendingReviewVersion - The version of the document that is currently under review.

$sel:documentFormat:DocumentDescription', documentDescription_documentFormat - The document format, either JSON or YAML.

$sel:tags:DocumentDescription', documentDescription_tags - The tags, or metadata, that have been applied to the document.

$sel:owner:DocumentDescription', documentDescription_owner - The Amazon Web Services user account that created the document.

$sel:attachmentsInformation:DocumentDescription', documentDescription_attachmentsInformation - Details about the document attachments, including names, locations, sizes, and so on.

$sel:reviewInformation:DocumentDescription', documentDescription_reviewInformation - Details about the review of a document.

$sel:description:DocumentDescription', documentDescription_description - A description of the document.

$sel:reviewStatus:DocumentDescription', documentDescription_reviewStatus - The current status of the review.

$sel:schemaVersion:DocumentDescription', documentDescription_schemaVersion - The schema version.

$sel:documentVersion:DocumentDescription', documentDescription_documentVersion - The document version.

$sel:displayName:DocumentDescription', documentDescription_displayName - The friendly name of the SSM document. This value can differ for each version of the document. If you want to update this value, see UpdateDocument.

$sel:hashType:DocumentDescription', documentDescription_hashType - The hash type of the document. Valid values include Sha256 or Sha1.

Sha1 hashes have been deprecated.

$sel:parameters:DocumentDescription', documentDescription_parameters - A description of the parameters for a document.

documentDescription_documentType :: Lens' DocumentDescription (Maybe DocumentType)

The type of document.

documentDescription_platformTypes :: Lens' DocumentDescription (Maybe [PlatformType])

The list of OS platforms compatible with this SSM document.

documentDescription_createdDate :: Lens' DocumentDescription (Maybe UTCTime)

The date when the document was created.

documentDescription_status :: Lens' DocumentDescription (Maybe DocumentStatus)

The status of the SSM document.

documentDescription_approvedVersion :: Lens' DocumentDescription (Maybe Text)

The version of the document currently approved for use in the organization.

documentDescription_defaultVersion :: Lens' DocumentDescription (Maybe Text)

The default version.

documentDescription_latestVersion :: Lens' DocumentDescription (Maybe Text)

The latest version of the document.

documentDescription_targetType :: Lens' DocumentDescription (Maybe Text)

The target type which defines the kinds of resources the document can run on. For example, /AWS::EC2::Instance. For a list of valid resource types, see Amazon Web Services resource and property types reference in the CloudFormation User Guide.

documentDescription_requires :: Lens' DocumentDescription (Maybe (NonEmpty DocumentRequires))

A list of SSM documents required by a document. For example, an ApplicationConfiguration document requires an ApplicationConfigurationSchema document.

documentDescription_sha1 :: Lens' DocumentDescription (Maybe Text)

The SHA1 hash of the document, which you can use for verification.

documentDescription_statusInformation :: Lens' DocumentDescription (Maybe Text)

A message returned by Amazon Web Services Systems Manager that explains the Status value. For example, a Failed status might be explained by the StatusInformation message, "The specified S3 bucket doesn't exist. Verify that the URL of the S3 bucket is correct."

documentDescription_author :: Lens' DocumentDescription (Maybe Text)

The user in your organization who created the document.

documentDescription_versionName :: Lens' DocumentDescription (Maybe Text)

The version of the artifact associated with the document.

documentDescription_hash :: Lens' DocumentDescription (Maybe Text)

The Sha256 or Sha1 hash created by the system when the document was created.

Sha1 hashes have been deprecated.

documentDescription_name :: Lens' DocumentDescription (Maybe Text)

The name of the SSM document.

documentDescription_pendingReviewVersion :: Lens' DocumentDescription (Maybe Text)

The version of the document that is currently under review.

documentDescription_documentFormat :: Lens' DocumentDescription (Maybe DocumentFormat)

The document format, either JSON or YAML.

documentDescription_tags :: Lens' DocumentDescription (Maybe [Tag])

The tags, or metadata, that have been applied to the document.

documentDescription_owner :: Lens' DocumentDescription (Maybe Text)

The Amazon Web Services user account that created the document.

documentDescription_attachmentsInformation :: Lens' DocumentDescription (Maybe [AttachmentInformation])

Details about the document attachments, including names, locations, sizes, and so on.

documentDescription_reviewInformation :: Lens' DocumentDescription (Maybe (NonEmpty ReviewInformation))

Details about the review of a document.

documentDescription_description :: Lens' DocumentDescription (Maybe Text)

A description of the document.

documentDescription_reviewStatus :: Lens' DocumentDescription (Maybe ReviewStatus)

The current status of the review.

documentDescription_schemaVersion :: Lens' DocumentDescription (Maybe Text)

The schema version.

documentDescription_documentVersion :: Lens' DocumentDescription (Maybe Text)

The document version.

documentDescription_displayName :: Lens' DocumentDescription (Maybe Text)

The friendly name of the SSM document. This value can differ for each version of the document. If you want to update this value, see UpdateDocument.

documentDescription_hashType :: Lens' DocumentDescription (Maybe DocumentHashType)

The hash type of the document. Valid values include Sha256 or Sha1.

Sha1 hashes have been deprecated.

documentDescription_parameters :: Lens' DocumentDescription (Maybe [DocumentParameter])

A description of the parameters for a document.