Network.AWS.SSM.Types.DocumentIdentifier

Description

Documentation

data DocumentIdentifier

Describes the name of a SSM document.

See: newDocumentIdentifier smart constructor.

Constructors

DocumentIdentifier'

Fields

documentType :: Maybe DocumentType

The document type.

platformTypes :: Maybe [PlatformType]

The operating system platform.

createdDate :: Maybe POSIX

The date the SSM document was created.

targetType :: Maybe Text

The target type which defines the kinds of resources the document can run on. For example, /AWS::EC2::Instance. For a list of valid resource types, see Amazon Web Services resource and property types reference in the CloudFormation User Guide.

requires :: Maybe (NonEmpty DocumentRequires)

A list of SSM documents required by a document. For example, an ApplicationConfiguration document requires an ApplicationConfigurationSchema document.

author :: Maybe Text

The user in your organization who created the document.

versionName :: Maybe Text

An optional field specifying the version of the artifact associated with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document, and can't be changed.

name :: Maybe Text

The name of the SSM document.

documentFormat :: Maybe DocumentFormat

The document format, either JSON or YAML.

tags :: Maybe [Tag]

The tags, or metadata, that have been applied to the document.

owner :: Maybe Text

The Amazon Web Services user account that created the document.

reviewStatus :: Maybe ReviewStatus

The current status of a document review.

schemaVersion :: Maybe Text

The schema version.

documentVersion :: Maybe Text

The document version.

displayName :: Maybe Text

An optional field where you can specify a friendly name for the SSM document. This value can differ for each version of the document. If you want to update this value, see UpdateDocument.

newDocumentIdentifier :: DocumentIdentifier

Create a value of DocumentIdentifier with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:documentType:DocumentIdentifier', documentIdentifier_documentType - The document type.

$sel:platformTypes:DocumentIdentifier', documentIdentifier_platformTypes - The operating system platform.

$sel:createdDate:DocumentIdentifier', documentIdentifier_createdDate - The date the SSM document was created.

$sel:targetType:DocumentIdentifier', documentIdentifier_targetType - The target type which defines the kinds of resources the document can run on. For example, /AWS::EC2::Instance. For a list of valid resource types, see Amazon Web Services resource and property types reference in the CloudFormation User Guide.

$sel:requires:DocumentIdentifier', documentIdentifier_requires - A list of SSM documents required by a document. For example, an ApplicationConfiguration document requires an ApplicationConfigurationSchema document.

$sel:author:DocumentIdentifier', documentIdentifier_author - The user in your organization who created the document.

$sel:versionName:DocumentIdentifier', documentIdentifier_versionName - An optional field specifying the version of the artifact associated with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document, and can't be changed.

$sel:name:DocumentIdentifier', documentIdentifier_name - The name of the SSM document.

$sel:documentFormat:DocumentIdentifier', documentIdentifier_documentFormat - The document format, either JSON or YAML.

$sel:tags:DocumentIdentifier', documentIdentifier_tags - The tags, or metadata, that have been applied to the document.

$sel:owner:DocumentIdentifier', documentIdentifier_owner - The Amazon Web Services user account that created the document.

$sel:reviewStatus:DocumentIdentifier', documentIdentifier_reviewStatus - The current status of a document review.

$sel:schemaVersion:DocumentIdentifier', documentIdentifier_schemaVersion - The schema version.

$sel:documentVersion:DocumentIdentifier', documentIdentifier_documentVersion - The document version.

$sel:displayName:DocumentIdentifier', documentIdentifier_displayName - An optional field where you can specify a friendly name for the SSM document. This value can differ for each version of the document. If you want to update this value, see UpdateDocument.

documentIdentifier_documentType :: Lens' DocumentIdentifier (Maybe DocumentType)

The document type.

documentIdentifier_platformTypes :: Lens' DocumentIdentifier (Maybe [PlatformType])

The operating system platform.

documentIdentifier_createdDate :: Lens' DocumentIdentifier (Maybe UTCTime)

The date the SSM document was created.

documentIdentifier_targetType :: Lens' DocumentIdentifier (Maybe Text)

The target type which defines the kinds of resources the document can run on. For example, /AWS::EC2::Instance. For a list of valid resource types, see Amazon Web Services resource and property types reference in the CloudFormation User Guide.

documentIdentifier_requires :: Lens' DocumentIdentifier (Maybe (NonEmpty DocumentRequires))

A list of SSM documents required by a document. For example, an ApplicationConfiguration document requires an ApplicationConfigurationSchema document.

documentIdentifier_author :: Lens' DocumentIdentifier (Maybe Text)

The user in your organization who created the document.

documentIdentifier_versionName :: Lens' DocumentIdentifier (Maybe Text)

An optional field specifying the version of the artifact associated with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document, and can't be changed.

documentIdentifier_name :: Lens' DocumentIdentifier (Maybe Text)

The name of the SSM document.

documentIdentifier_documentFormat :: Lens' DocumentIdentifier (Maybe DocumentFormat)

The document format, either JSON or YAML.

documentIdentifier_tags :: Lens' DocumentIdentifier (Maybe [Tag])

The tags, or metadata, that have been applied to the document.

documentIdentifier_owner :: Lens' DocumentIdentifier (Maybe Text)

The Amazon Web Services user account that created the document.

documentIdentifier_reviewStatus :: Lens' DocumentIdentifier (Maybe ReviewStatus)

The current status of a document review.

documentIdentifier_schemaVersion :: Lens' DocumentIdentifier (Maybe Text)

The schema version.

documentIdentifier_documentVersion :: Lens' DocumentIdentifier (Maybe Text)

The document version.

documentIdentifier_displayName :: Lens' DocumentIdentifier (Maybe Text)

An optional field where you can specify a friendly name for the SSM document. This value can differ for each version of the document. If you want to update this value, see UpdateDocument.