Network.AWS.SSM.Types.InstanceAssociationStatusInfo

Description

Documentation

data InstanceAssociationStatusInfo

Status information about the instance association.

See: newInstanceAssociationStatusInfo smart constructor.

Constructors

InstanceAssociationStatusInfo'

Fields

associationId :: Maybe Text

The association ID.

instanceId :: Maybe Text

The instance ID where the association was created.

detailedStatus :: Maybe Text

Detailed status information about the instance association.

status :: Maybe Text

Status information about the instance association.

outputUrl :: Maybe InstanceAssociationOutputUrl

A URL for an S3 bucket where you want to store the results of this request.

executionSummary :: Maybe Text

Summary information about association execution.

name :: Maybe Text

The name of the association.

errorCode :: Maybe Text

An error code returned by the request to create the association.

documentVersion :: Maybe Text

The association document versions.

associationVersion :: Maybe Text

The version of the association applied to the instance.

executionDate :: Maybe POSIX

The date the instance association ran.

associationName :: Maybe Text

The name of the association applied to the instance.

newInstanceAssociationStatusInfo :: InstanceAssociationStatusInfo

Create a value of InstanceAssociationStatusInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associationId:InstanceAssociationStatusInfo', instanceAssociationStatusInfo_associationId - The association ID.

$sel:instanceId:InstanceAssociationStatusInfo', instanceAssociationStatusInfo_instanceId - The instance ID where the association was created.

$sel:detailedStatus:InstanceAssociationStatusInfo', instanceAssociationStatusInfo_detailedStatus - Detailed status information about the instance association.

$sel:status:InstanceAssociationStatusInfo', instanceAssociationStatusInfo_status - Status information about the instance association.

$sel:outputUrl:InstanceAssociationStatusInfo', instanceAssociationStatusInfo_outputUrl - A URL for an S3 bucket where you want to store the results of this request.

$sel:executionSummary:InstanceAssociationStatusInfo', instanceAssociationStatusInfo_executionSummary - Summary information about association execution.

$sel:name:InstanceAssociationStatusInfo', instanceAssociationStatusInfo_name - The name of the association.

$sel:errorCode:InstanceAssociationStatusInfo', instanceAssociationStatusInfo_errorCode - An error code returned by the request to create the association.

$sel:documentVersion:InstanceAssociationStatusInfo', instanceAssociationStatusInfo_documentVersion - The association document versions.

$sel:associationVersion:InstanceAssociationStatusInfo', instanceAssociationStatusInfo_associationVersion - The version of the association applied to the instance.

$sel:executionDate:InstanceAssociationStatusInfo', instanceAssociationStatusInfo_executionDate - The date the instance association ran.

$sel:associationName:InstanceAssociationStatusInfo', instanceAssociationStatusInfo_associationName - The name of the association applied to the instance.

instanceAssociationStatusInfo_associationId :: Lens' InstanceAssociationStatusInfo (Maybe Text)

The association ID.

instanceAssociationStatusInfo_instanceId :: Lens' InstanceAssociationStatusInfo (Maybe Text)

The instance ID where the association was created.

instanceAssociationStatusInfo_detailedStatus :: Lens' InstanceAssociationStatusInfo (Maybe Text)

Detailed status information about the instance association.

instanceAssociationStatusInfo_status :: Lens' InstanceAssociationStatusInfo (Maybe Text)

Status information about the instance association.

instanceAssociationStatusInfo_outputUrl :: Lens' InstanceAssociationStatusInfo (Maybe InstanceAssociationOutputUrl)

A URL for an S3 bucket where you want to store the results of this request.

instanceAssociationStatusInfo_executionSummary :: Lens' InstanceAssociationStatusInfo (Maybe Text)

Summary information about association execution.

instanceAssociationStatusInfo_name :: Lens' InstanceAssociationStatusInfo (Maybe Text)

The name of the association.

instanceAssociationStatusInfo_errorCode :: Lens' InstanceAssociationStatusInfo (Maybe Text)

An error code returned by the request to create the association.

instanceAssociationStatusInfo_documentVersion :: Lens' InstanceAssociationStatusInfo (Maybe Text)

The association document versions.

instanceAssociationStatusInfo_associationVersion :: Lens' InstanceAssociationStatusInfo (Maybe Text)

The version of the association applied to the instance.

instanceAssociationStatusInfo_executionDate :: Lens' InstanceAssociationStatusInfo (Maybe UTCTime)

The date the instance association ran.

instanceAssociationStatusInfo_associationName :: Lens' InstanceAssociationStatusInfo (Maybe Text)

The name of the association applied to the instance.