Network.AWS.SSM.Types.InstanceInformation

Description

Documentation

data InstanceInformation

Describes a filter for a specific list of instances.

See: newInstanceInformation smart constructor.

Constructors

InstanceInformation'

Fields

instanceId :: Maybe Text

The instance ID.

pingStatus :: Maybe PingStatus

Connection status of SSM Agent.

The status Inactive has been deprecated and is no longer in use.

iamRole :: Maybe Text

The Identity and Access Management (IAM) role assigned to the on-premises Systems Manager managed instance. This call doesn't return the IAM role for Amazon Elastic Compute Cloud (Amazon EC2) instances. To retrieve the IAM role for an EC2 instance, use the Amazon EC2 DescribeInstances operation. For information, see DescribeInstances in the Amazon EC2 API Reference or describe-instances in the Amazon Web Services CLI Command Reference.

activationId :: Maybe Text

The activation ID created by Amazon Web Services Systems Manager when the server or virtual machine (VM) was registered.

lastSuccessfulAssociationExecutionDate :: Maybe POSIX

The last date the association was successfully run.

agentVersion :: Maybe Text

The version of SSM Agent running on your Linux instance.

lastPingDateTime :: Maybe POSIX

The date and time when the agent last pinged the Systems Manager service.

platformVersion :: Maybe Text

The version of the OS platform running on your instance.

lastAssociationExecutionDate :: Maybe POSIX

The date the association was last run.

resourceType :: Maybe ResourceType

The type of instance. Instances are either EC2 instances or managed instances.

associationOverview :: Maybe InstanceAggregatedAssociationOverview

Information about the association.

name :: Maybe Text

The name assigned to an on-premises server or virtual machine (VM) when it is activated as a Systems Manager managed instance. The name is specified as the DefaultInstanceName property using the CreateActivation command. It is applied to the managed instance by specifying the Activation Code and Activation ID when you install SSM Agent on the instance, as explained in Install SSM Agent for a hybrid environment (Linux) and Install SSM Agent for a hybrid environment (Windows). To retrieve the Name tag of an EC2 instance, use the Amazon EC2 DescribeInstances operation. For information, see DescribeInstances in the Amazon EC2 API Reference or describe-instances in the Amazon Web Services CLI Command Reference.

iPAddress :: Maybe Text

The IP address of the managed instance.

platformType :: Maybe PlatformType

The operating system platform type.

isLatestVersion :: Maybe Bool

Indicates whether the latest version of SSM Agent is running on your Linux Managed Instance. This field doesn't indicate whether or not the latest version is installed on Windows managed instances, because some older versions of Windows Server use the EC2Config service to process Systems Manager requests.

platformName :: Maybe Text

The name of the operating system platform running on your instance.

computerName :: Maybe Text

The fully qualified host name of the managed instance.

associationStatus :: Maybe Text

The status of the association.

registrationDate :: Maybe POSIX

The date the server or VM was registered with Amazon Web Services as a managed instance.

newInstanceInformation :: InstanceInformation

Create a value of InstanceInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:InstanceInformation', instanceInformation_instanceId - The instance ID.

$sel:pingStatus:InstanceInformation', instanceInformation_pingStatus - Connection status of SSM Agent.

The status Inactive has been deprecated and is no longer in use.

$sel:iamRole:InstanceInformation', instanceInformation_iamRole - The Identity and Access Management (IAM) role assigned to the on-premises Systems Manager managed instance. This call doesn't return the IAM role for Amazon Elastic Compute Cloud (Amazon EC2) instances. To retrieve the IAM role for an EC2 instance, use the Amazon EC2 DescribeInstances operation. For information, see DescribeInstances in the Amazon EC2 API Reference or describe-instances in the Amazon Web Services CLI Command Reference.

$sel:activationId:InstanceInformation', instanceInformation_activationId - The activation ID created by Amazon Web Services Systems Manager when the server or virtual machine (VM) was registered.

$sel:lastSuccessfulAssociationExecutionDate:InstanceInformation', instanceInformation_lastSuccessfulAssociationExecutionDate - The last date the association was successfully run.

$sel:agentVersion:InstanceInformation', instanceInformation_agentVersion - The version of SSM Agent running on your Linux instance.

$sel:lastPingDateTime:InstanceInformation', instanceInformation_lastPingDateTime - The date and time when the agent last pinged the Systems Manager service.

$sel:platformVersion:InstanceInformation', instanceInformation_platformVersion - The version of the OS platform running on your instance.

$sel:lastAssociationExecutionDate:InstanceInformation', instanceInformation_lastAssociationExecutionDate - The date the association was last run.

$sel:resourceType:InstanceInformation', instanceInformation_resourceType - The type of instance. Instances are either EC2 instances or managed instances.

$sel:associationOverview:InstanceInformation', instanceInformation_associationOverview - Information about the association.

$sel:name:InstanceInformation', instanceInformation_name - The name assigned to an on-premises server or virtual machine (VM) when it is activated as a Systems Manager managed instance. The name is specified as the DefaultInstanceName property using the CreateActivation command. It is applied to the managed instance by specifying the Activation Code and Activation ID when you install SSM Agent on the instance, as explained in Install SSM Agent for a hybrid environment (Linux) and Install SSM Agent for a hybrid environment (Windows). To retrieve the Name tag of an EC2 instance, use the Amazon EC2 DescribeInstances operation. For information, see DescribeInstances in the Amazon EC2 API Reference or describe-instances in the Amazon Web Services CLI Command Reference.

$sel:iPAddress:InstanceInformation', instanceInformation_iPAddress - The IP address of the managed instance.

$sel:platformType:InstanceInformation', instanceInformation_platformType - The operating system platform type.

$sel:isLatestVersion:InstanceInformation', instanceInformation_isLatestVersion - Indicates whether the latest version of SSM Agent is running on your Linux Managed Instance. This field doesn't indicate whether or not the latest version is installed on Windows managed instances, because some older versions of Windows Server use the EC2Config service to process Systems Manager requests.

$sel:platformName:InstanceInformation', instanceInformation_platformName - The name of the operating system platform running on your instance.

$sel:computerName:InstanceInformation', instanceInformation_computerName - The fully qualified host name of the managed instance.

$sel:associationStatus:InstanceInformation', instanceInformation_associationStatus - The status of the association.

$sel:registrationDate:InstanceInformation', instanceInformation_registrationDate - The date the server or VM was registered with Amazon Web Services as a managed instance.

instanceInformation_instanceId :: Lens' InstanceInformation (Maybe Text)

The instance ID.

instanceInformation_pingStatus :: Lens' InstanceInformation (Maybe PingStatus)

Connection status of SSM Agent.

The status Inactive has been deprecated and is no longer in use.

instanceInformation_iamRole :: Lens' InstanceInformation (Maybe Text)

The Identity and Access Management (IAM) role assigned to the on-premises Systems Manager managed instance. This call doesn't return the IAM role for Amazon Elastic Compute Cloud (Amazon EC2) instances. To retrieve the IAM role for an EC2 instance, use the Amazon EC2 DescribeInstances operation. For information, see DescribeInstances in the Amazon EC2 API Reference or describe-instances in the Amazon Web Services CLI Command Reference.

instanceInformation_activationId :: Lens' InstanceInformation (Maybe Text)

The activation ID created by Amazon Web Services Systems Manager when the server or virtual machine (VM) was registered.

instanceInformation_lastSuccessfulAssociationExecutionDate :: Lens' InstanceInformation (Maybe UTCTime)

The last date the association was successfully run.

instanceInformation_agentVersion :: Lens' InstanceInformation (Maybe Text)

The version of SSM Agent running on your Linux instance.

instanceInformation_lastPingDateTime :: Lens' InstanceInformation (Maybe UTCTime)

The date and time when the agent last pinged the Systems Manager service.

instanceInformation_platformVersion :: Lens' InstanceInformation (Maybe Text)

The version of the OS platform running on your instance.

instanceInformation_lastAssociationExecutionDate :: Lens' InstanceInformation (Maybe UTCTime)

The date the association was last run.

instanceInformation_resourceType :: Lens' InstanceInformation (Maybe ResourceType)

The type of instance. Instances are either EC2 instances or managed instances.

instanceInformation_associationOverview :: Lens' InstanceInformation (Maybe InstanceAggregatedAssociationOverview)

Information about the association.

instanceInformation_name :: Lens' InstanceInformation (Maybe Text)

The name assigned to an on-premises server or virtual machine (VM) when it is activated as a Systems Manager managed instance. The name is specified as the DefaultInstanceName property using the CreateActivation command. It is applied to the managed instance by specifying the Activation Code and Activation ID when you install SSM Agent on the instance, as explained in Install SSM Agent for a hybrid environment (Linux) and Install SSM Agent for a hybrid environment (Windows). To retrieve the Name tag of an EC2 instance, use the Amazon EC2 DescribeInstances operation. For information, see DescribeInstances in the Amazon EC2 API Reference or describe-instances in the Amazon Web Services CLI Command Reference.

instanceInformation_iPAddress :: Lens' InstanceInformation (Maybe Text)

The IP address of the managed instance.

instanceInformation_platformType :: Lens' InstanceInformation (Maybe PlatformType)

The operating system platform type.

instanceInformation_isLatestVersion :: Lens' InstanceInformation (Maybe Bool)

Indicates whether the latest version of SSM Agent is running on your Linux Managed Instance. This field doesn't indicate whether or not the latest version is installed on Windows managed instances, because some older versions of Windows Server use the EC2Config service to process Systems Manager requests.

instanceInformation_platformName :: Lens' InstanceInformation (Maybe Text)

The name of the operating system platform running on your instance.

instanceInformation_computerName :: Lens' InstanceInformation (Maybe Text)

The fully qualified host name of the managed instance.

instanceInformation_associationStatus :: Lens' InstanceInformation (Maybe Text)

The status of the association.

instanceInformation_registrationDate :: Lens' InstanceInformation (Maybe UTCTime)

The date the server or VM was registered with Amazon Web Services as a managed instance.