Network.AWS.SSM.Types.InventoryDeletionStatusItem

Description

Documentation

data InventoryDeletionStatusItem

Status information returned by the DeleteInventory operation.

See: newInventoryDeletionStatusItem smart constructor.

Constructors

InventoryDeletionStatusItem'

Fields

typeName :: Maybe Text

The name of the inventory data type.

lastStatusMessage :: Maybe Text

Information about the status.

lastStatusUpdateTime :: Maybe POSIX

The UTC timestamp of when the last status report.

deletionStartTime :: Maybe POSIX

The UTC timestamp when the delete operation started.

deletionId :: Maybe Text

The deletion ID returned by the DeleteInventory operation.

lastStatus :: Maybe InventoryDeletionStatus

The status of the operation. Possible values are InProgress and Complete.

deletionSummary :: Maybe InventoryDeletionSummary

Information about the delete operation. For more information about this summary, see Understanding the delete inventory summary in the Amazon Web Services Systems Manager User Guide.

newInventoryDeletionStatusItem :: InventoryDeletionStatusItem

Create a value of InventoryDeletionStatusItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeName:InventoryDeletionStatusItem', inventoryDeletionStatusItem_typeName - The name of the inventory data type.

$sel:lastStatusMessage:InventoryDeletionStatusItem', inventoryDeletionStatusItem_lastStatusMessage - Information about the status.

$sel:lastStatusUpdateTime:InventoryDeletionStatusItem', inventoryDeletionStatusItem_lastStatusUpdateTime - The UTC timestamp of when the last status report.

$sel:deletionStartTime:InventoryDeletionStatusItem', inventoryDeletionStatusItem_deletionStartTime - The UTC timestamp when the delete operation started.

$sel:deletionId:InventoryDeletionStatusItem', inventoryDeletionStatusItem_deletionId - The deletion ID returned by the DeleteInventory operation.

$sel:lastStatus:InventoryDeletionStatusItem', inventoryDeletionStatusItem_lastStatus - The status of the operation. Possible values are InProgress and Complete.

$sel:deletionSummary:InventoryDeletionStatusItem', inventoryDeletionStatusItem_deletionSummary - Information about the delete operation. For more information about this summary, see Understanding the delete inventory summary in the Amazon Web Services Systems Manager User Guide.

inventoryDeletionStatusItem_typeName :: Lens' InventoryDeletionStatusItem (Maybe Text)

The name of the inventory data type.

inventoryDeletionStatusItem_lastStatusMessage :: Lens' InventoryDeletionStatusItem (Maybe Text)

Information about the status.

inventoryDeletionStatusItem_lastStatusUpdateTime :: Lens' InventoryDeletionStatusItem (Maybe UTCTime)

The UTC timestamp of when the last status report.

inventoryDeletionStatusItem_deletionStartTime :: Lens' InventoryDeletionStatusItem (Maybe UTCTime)

The UTC timestamp when the delete operation started.

inventoryDeletionStatusItem_deletionId :: Lens' InventoryDeletionStatusItem (Maybe Text)

The deletion ID returned by the DeleteInventory operation.

inventoryDeletionStatusItem_lastStatus :: Lens' InventoryDeletionStatusItem (Maybe InventoryDeletionStatus)

The status of the operation. Possible values are InProgress and Complete.

inventoryDeletionStatusItem_deletionSummary :: Lens' InventoryDeletionStatusItem (Maybe InventoryDeletionSummary)

Information about the delete operation. For more information about this summary, see Understanding the delete inventory summary in the Amazon Web Services Systems Manager User Guide.