Network.AWS.SSM.Types.InventoryResultEntity

Description

Documentation

data InventoryResultEntity

Inventory query results.

See: newInventoryResultEntity smart constructor.

Constructors

InventoryResultEntity'

Fields

data' :: Maybe (HashMap Text InventoryResultItem)

The data section in the inventory result entity JSON.

id :: Maybe Text

ID of the inventory result entity. For example, for managed instance inventory the result will be the managed instance ID. For EC2 instance inventory, the result will be the instance ID.

newInventoryResultEntity :: InventoryResultEntity

Create a value of InventoryResultEntity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:data':InventoryResultEntity', inventoryResultEntity_data - The data section in the inventory result entity JSON.

$sel:id:InventoryResultEntity', inventoryResultEntity_id - ID of the inventory result entity. For example, for managed instance inventory the result will be the managed instance ID. For EC2 instance inventory, the result will be the instance ID.

inventoryResultEntity_data :: Lens' InventoryResultEntity (Maybe (HashMap Text InventoryResultItem))

The data section in the inventory result entity JSON.

inventoryResultEntity_id :: Lens' InventoryResultEntity (Maybe Text)

ID of the inventory result entity. For example, for managed instance inventory the result will be the managed instance ID. For EC2 instance inventory, the result will be the instance ID.