Network.AWS.SSM.Types.InventoryResultItem

Description

Documentation

data InventoryResultItem

The inventory result item.

See: newInventoryResultItem smart constructor.

Constructors

InventoryResultItem'

Fields

contentHash :: Maybe Text

MD5 hash of the inventory item type contents. The content hash is used to determine whether to update inventory information. The PutInventory API doesn't update the inventory item type contents if the MD5 hash hasn't changed since last update.

captureTime :: Maybe Text

The time inventory item data was captured.

typeName :: Text

The name of the inventory result item type.

schemaVersion :: Text

The schema version for the inventory result item/

content :: [HashMap Text Text]

Contains all the inventory data of the item type. Results include attribute names and values.

newInventoryResultItem

Arguments

:: Text
$sel:typeName:InventoryResultItem'

-> Text
$sel:schemaVersion:InventoryResultItem'

-> InventoryResultItem

Create a value of InventoryResultItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentHash:InventoryResultItem', inventoryResultItem_contentHash - MD5 hash of the inventory item type contents. The content hash is used to determine whether to update inventory information. The PutInventory API doesn't update the inventory item type contents if the MD5 hash hasn't changed since last update.

$sel:captureTime:InventoryResultItem', inventoryResultItem_captureTime - The time inventory item data was captured.

$sel:typeName:InventoryResultItem', inventoryResultItem_typeName - The name of the inventory result item type.

$sel:schemaVersion:InventoryResultItem', inventoryResultItem_schemaVersion - The schema version for the inventory result item/

$sel:content:InventoryResultItem', inventoryResultItem_content - Contains all the inventory data of the item type. Results include attribute names and values.

inventoryResultItem_contentHash :: Lens' InventoryResultItem (Maybe Text)

MD5 hash of the inventory item type contents. The content hash is used to determine whether to update inventory information. The PutInventory API doesn't update the inventory item type contents if the MD5 hash hasn't changed since last update.

inventoryResultItem_captureTime :: Lens' InventoryResultItem (Maybe Text)

The time inventory item data was captured.

inventoryResultItem_typeName :: Lens' InventoryResultItem Text

The name of the inventory result item type.

inventoryResultItem_schemaVersion :: Lens' InventoryResultItem Text

The schema version for the inventory result item/

inventoryResultItem_content :: Lens' InventoryResultItem [HashMap Text Text]

Contains all the inventory data of the item type. Results include attribute names and values.