Network.AWS.SSM.Types.MaintenanceWindowExecutionTaskIdentity

Description

Documentation

data MaintenanceWindowExecutionTaskIdentity

Information about a task execution performed as part of a maintenance window execution.

See: newMaintenanceWindowExecutionTaskIdentity smart constructor.

Constructors

MaintenanceWindowExecutionTaskIdentity'

Fields

status :: Maybe MaintenanceWindowExecutionStatus

The status of the task execution.

taskExecutionId :: Maybe Text

The ID of the specific task execution in the maintenance window execution.

startTime :: Maybe POSIX

The time the task execution started.

taskType :: Maybe MaintenanceWindowTaskType

The type of task that ran.

taskArn :: Maybe Text

The Amazon Resource Name (ARN) of the task that ran.

windowExecutionId :: Maybe Text

The ID of the maintenance window execution that ran the task.

statusDetails :: Maybe Text

The details explaining the status of the task execution. Not available for all status values.

endTime :: Maybe POSIX

The time the task execution finished.

newMaintenanceWindowExecutionTaskIdentity :: MaintenanceWindowExecutionTaskIdentity

Create a value of MaintenanceWindowExecutionTaskIdentity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:MaintenanceWindowExecutionTaskIdentity', maintenanceWindowExecutionTaskIdentity_status - The status of the task execution.

$sel:taskExecutionId:MaintenanceWindowExecutionTaskIdentity', maintenanceWindowExecutionTaskIdentity_taskExecutionId - The ID of the specific task execution in the maintenance window execution.

$sel:startTime:MaintenanceWindowExecutionTaskIdentity', maintenanceWindowExecutionTaskIdentity_startTime - The time the task execution started.

$sel:taskType:MaintenanceWindowExecutionTaskIdentity', maintenanceWindowExecutionTaskIdentity_taskType - The type of task that ran.

$sel:taskArn:MaintenanceWindowExecutionTaskIdentity', maintenanceWindowExecutionTaskIdentity_taskArn - The Amazon Resource Name (ARN) of the task that ran.

$sel:windowExecutionId:MaintenanceWindowExecutionTaskIdentity', maintenanceWindowExecutionTaskIdentity_windowExecutionId - The ID of the maintenance window execution that ran the task.

$sel:statusDetails:MaintenanceWindowExecutionTaskIdentity', maintenanceWindowExecutionTaskIdentity_statusDetails - The details explaining the status of the task execution. Not available for all status values.

$sel:endTime:MaintenanceWindowExecutionTaskIdentity', maintenanceWindowExecutionTaskIdentity_endTime - The time the task execution finished.

maintenanceWindowExecutionTaskIdentity_status :: Lens' MaintenanceWindowExecutionTaskIdentity (Maybe MaintenanceWindowExecutionStatus)

The status of the task execution.

maintenanceWindowExecutionTaskIdentity_taskExecutionId :: Lens' MaintenanceWindowExecutionTaskIdentity (Maybe Text)

The ID of the specific task execution in the maintenance window execution.

maintenanceWindowExecutionTaskIdentity_startTime :: Lens' MaintenanceWindowExecutionTaskIdentity (Maybe UTCTime)

The time the task execution started.

maintenanceWindowExecutionTaskIdentity_taskType :: Lens' MaintenanceWindowExecutionTaskIdentity (Maybe MaintenanceWindowTaskType)

The type of task that ran.

maintenanceWindowExecutionTaskIdentity_taskArn :: Lens' MaintenanceWindowExecutionTaskIdentity (Maybe Text)

The Amazon Resource Name (ARN) of the task that ran.

maintenanceWindowExecutionTaskIdentity_windowExecutionId :: Lens' MaintenanceWindowExecutionTaskIdentity (Maybe Text)

The ID of the maintenance window execution that ran the task.

maintenanceWindowExecutionTaskIdentity_statusDetails :: Lens' MaintenanceWindowExecutionTaskIdentity (Maybe Text)

The details explaining the status of the task execution. Not available for all status values.

maintenanceWindowExecutionTaskIdentity_endTime :: Lens' MaintenanceWindowExecutionTaskIdentity (Maybe UTCTime)

The time the task execution finished.