Network.AWS.SSM.Types.MaintenanceWindowExecutionTaskInvocationIdentity

Description

Documentation

data MaintenanceWindowExecutionTaskInvocationIdentity

Describes the information about a task invocation for a particular target as part of a task execution performed as part of a maintenance window execution.

See: newMaintenanceWindowExecutionTaskInvocationIdentity smart constructor.

Constructors

MaintenanceWindowExecutionTaskInvocationIdentity'

Fields

status :: Maybe MaintenanceWindowExecutionStatus

The status of the task invocation.

statusDetails :: Maybe Text

The details explaining the status of the task invocation. Not available for all status values.

windowTargetId :: Maybe Text

The ID of the target definition in this maintenance window the invocation was performed for.

startTime :: Maybe POSIX

The time the invocation started.

endTime :: Maybe POSIX

The time the invocation finished.

executionId :: Maybe Text

The ID of the action performed in the service that actually handled the task invocation. If the task type is RUN_COMMAND, this value is the command ID.

windowExecutionId :: Maybe Text

The ID of the maintenance window execution that ran the task.

invocationId :: Maybe Text

The ID of the task invocation.

ownerInformation :: Maybe (Sensitive Text)

User-provided value that was specified when the target was registered with the maintenance window. This was also included in any Amazon CloudWatch Events events raised during the task invocation.

taskType :: Maybe MaintenanceWindowTaskType

The task type.

parameters :: Maybe (Sensitive Text)

The parameters that were provided for the invocation when it was run.

taskExecutionId :: Maybe Text

The ID of the specific task execution in the maintenance window execution.

newMaintenanceWindowExecutionTaskInvocationIdentity :: MaintenanceWindowExecutionTaskInvocationIdentity

Create a value of MaintenanceWindowExecutionTaskInvocationIdentity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:MaintenanceWindowExecutionTaskInvocationIdentity', maintenanceWindowExecutionTaskInvocationIdentity_status - The status of the task invocation.

$sel:statusDetails:MaintenanceWindowExecutionTaskInvocationIdentity', maintenanceWindowExecutionTaskInvocationIdentity_statusDetails - The details explaining the status of the task invocation. Not available for all status values.

$sel:windowTargetId:MaintenanceWindowExecutionTaskInvocationIdentity', maintenanceWindowExecutionTaskInvocationIdentity_windowTargetId - The ID of the target definition in this maintenance window the invocation was performed for.

$sel:startTime:MaintenanceWindowExecutionTaskInvocationIdentity', maintenanceWindowExecutionTaskInvocationIdentity_startTime - The time the invocation started.

$sel:endTime:MaintenanceWindowExecutionTaskInvocationIdentity', maintenanceWindowExecutionTaskInvocationIdentity_endTime - The time the invocation finished.

$sel:executionId:MaintenanceWindowExecutionTaskInvocationIdentity', maintenanceWindowExecutionTaskInvocationIdentity_executionId - The ID of the action performed in the service that actually handled the task invocation. If the task type is RUN_COMMAND, this value is the command ID.

$sel:windowExecutionId:MaintenanceWindowExecutionTaskInvocationIdentity', maintenanceWindowExecutionTaskInvocationIdentity_windowExecutionId - The ID of the maintenance window execution that ran the task.

$sel:invocationId:MaintenanceWindowExecutionTaskInvocationIdentity', maintenanceWindowExecutionTaskInvocationIdentity_invocationId - The ID of the task invocation.

$sel:ownerInformation:MaintenanceWindowExecutionTaskInvocationIdentity', maintenanceWindowExecutionTaskInvocationIdentity_ownerInformation - User-provided value that was specified when the target was registered with the maintenance window. This was also included in any Amazon CloudWatch Events events raised during the task invocation.

$sel:taskType:MaintenanceWindowExecutionTaskInvocationIdentity', maintenanceWindowExecutionTaskInvocationIdentity_taskType - The task type.

$sel:parameters:MaintenanceWindowExecutionTaskInvocationIdentity', maintenanceWindowExecutionTaskInvocationIdentity_parameters - The parameters that were provided for the invocation when it was run.

$sel:taskExecutionId:MaintenanceWindowExecutionTaskInvocationIdentity', maintenanceWindowExecutionTaskInvocationIdentity_taskExecutionId - The ID of the specific task execution in the maintenance window execution.

maintenanceWindowExecutionTaskInvocationIdentity_status :: Lens' MaintenanceWindowExecutionTaskInvocationIdentity (Maybe MaintenanceWindowExecutionStatus)

The status of the task invocation.

maintenanceWindowExecutionTaskInvocationIdentity_statusDetails :: Lens' MaintenanceWindowExecutionTaskInvocationIdentity (Maybe Text)

The details explaining the status of the task invocation. Not available for all status values.

maintenanceWindowExecutionTaskInvocationIdentity_windowTargetId :: Lens' MaintenanceWindowExecutionTaskInvocationIdentity (Maybe Text)

The ID of the target definition in this maintenance window the invocation was performed for.

maintenanceWindowExecutionTaskInvocationIdentity_startTime :: Lens' MaintenanceWindowExecutionTaskInvocationIdentity (Maybe UTCTime)

The time the invocation started.

maintenanceWindowExecutionTaskInvocationIdentity_endTime :: Lens' MaintenanceWindowExecutionTaskInvocationIdentity (Maybe UTCTime)

The time the invocation finished.

maintenanceWindowExecutionTaskInvocationIdentity_executionId :: Lens' MaintenanceWindowExecutionTaskInvocationIdentity (Maybe Text)

The ID of the action performed in the service that actually handled the task invocation. If the task type is RUN_COMMAND, this value is the command ID.

maintenanceWindowExecutionTaskInvocationIdentity_windowExecutionId :: Lens' MaintenanceWindowExecutionTaskInvocationIdentity (Maybe Text)

The ID of the maintenance window execution that ran the task.

maintenanceWindowExecutionTaskInvocationIdentity_invocationId :: Lens' MaintenanceWindowExecutionTaskInvocationIdentity (Maybe Text)

The ID of the task invocation.

maintenanceWindowExecutionTaskInvocationIdentity_ownerInformation :: Lens' MaintenanceWindowExecutionTaskInvocationIdentity (Maybe Text)

User-provided value that was specified when the target was registered with the maintenance window. This was also included in any Amazon CloudWatch Events events raised during the task invocation.

maintenanceWindowExecutionTaskInvocationIdentity_taskType :: Lens' MaintenanceWindowExecutionTaskInvocationIdentity (Maybe MaintenanceWindowTaskType)

The task type.

maintenanceWindowExecutionTaskInvocationIdentity_parameters :: Lens' MaintenanceWindowExecutionTaskInvocationIdentity (Maybe Text)

The parameters that were provided for the invocation when it was run.

maintenanceWindowExecutionTaskInvocationIdentity_taskExecutionId :: Lens' MaintenanceWindowExecutionTaskInvocationIdentity (Maybe Text)

The ID of the specific task execution in the maintenance window execution.