Network.AWS.SSM.Types.MaintenanceWindowRunCommandParameters

Description

Documentation

data MaintenanceWindowRunCommandParameters

The parameters for a RUN_COMMAND task type.

For information about specifying and updating task parameters, see RegisterTaskWithMaintenanceWindow and UpdateMaintenanceWindowTask.

LoggingInfo has been deprecated. To specify an Amazon Simple Storage Service (Amazon S3) bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Amazon Web Services Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

For RUN_COMMAND tasks, Systems Manager uses specified values for TaskParameters and LoggingInfo only if no values are specified for TaskInvocationParameters.

See: newMaintenanceWindowRunCommandParameters smart constructor.

Constructors

MaintenanceWindowRunCommandParameters'

Fields

notificationConfig :: Maybe NotificationConfig

Configurations for sending notifications about command status changes on a per-instance basis.

serviceRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) service role to use to publish Amazon Simple Notification Service (Amazon SNS) notifications for maintenance window Run Command tasks.

outputS3BucketName :: Maybe Text

The name of the Amazon Simple Storage Service (Amazon S3) bucket.

comment :: Maybe Text

Information about the commands to run.

documentHash :: Maybe Text

The SHA-256 or SHA-1 hash created by the system when the document was created. SHA-1 hashes have been deprecated.

timeoutSeconds :: Maybe Natural

If this time is reached and the command hasn't already started running, it doesn't run.

outputS3KeyPrefix :: Maybe Text

The S3 bucket subfolder.

cloudWatchOutputConfig :: Maybe CloudWatchOutputConfig

documentVersion :: Maybe Text

The Amazon Web Services Systems Manager document (SSM document) version to use in the request. You can specify $DEFAULT, $LATEST, or a specific version number. If you run commands by using the Amazon Web Services CLI, then you must escape the first two options by using a backslash. If you specify a version number, then you don't need to use the backslash. For example:

--document-version "\$DEFAULT"

--document-version "\$LATEST"

--document-version "3"

documentHashType :: Maybe DocumentHashType

SHA-256 or SHA-1. SHA-1 hashes have been deprecated.

parameters :: Maybe (HashMap Text [Text])

The parameters for the RUN_COMMAND task execution.

newMaintenanceWindowRunCommandParameters :: MaintenanceWindowRunCommandParameters

Create a value of MaintenanceWindowRunCommandParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notificationConfig:MaintenanceWindowRunCommandParameters', maintenanceWindowRunCommandParameters_notificationConfig - Configurations for sending notifications about command status changes on a per-instance basis.

$sel:serviceRoleArn:MaintenanceWindowRunCommandParameters', maintenanceWindowRunCommandParameters_serviceRoleArn - The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) service role to use to publish Amazon Simple Notification Service (Amazon SNS) notifications for maintenance window Run Command tasks.

$sel:outputS3BucketName:MaintenanceWindowRunCommandParameters', maintenanceWindowRunCommandParameters_outputS3BucketName - The name of the Amazon Simple Storage Service (Amazon S3) bucket.

$sel:comment:MaintenanceWindowRunCommandParameters', maintenanceWindowRunCommandParameters_comment - Information about the commands to run.

$sel:documentHash:MaintenanceWindowRunCommandParameters', maintenanceWindowRunCommandParameters_documentHash - The SHA-256 or SHA-1 hash created by the system when the document was created. SHA-1 hashes have been deprecated.

$sel:timeoutSeconds:MaintenanceWindowRunCommandParameters', maintenanceWindowRunCommandParameters_timeoutSeconds - If this time is reached and the command hasn't already started running, it doesn't run.

$sel:outputS3KeyPrefix:MaintenanceWindowRunCommandParameters', maintenanceWindowRunCommandParameters_outputS3KeyPrefix - The S3 bucket subfolder.

$sel:cloudWatchOutputConfig:MaintenanceWindowRunCommandParameters', maintenanceWindowRunCommandParameters_cloudWatchOutputConfig - Undocumented member.

$sel:documentVersion:MaintenanceWindowRunCommandParameters', maintenanceWindowRunCommandParameters_documentVersion - The Amazon Web Services Systems Manager document (SSM document) version to use in the request. You can specify $DEFAULT, $LATEST, or a specific version number. If you run commands by using the Amazon Web Services CLI, then you must escape the first two options by using a backslash. If you specify a version number, then you don't need to use the backslash. For example:

--document-version "\$DEFAULT"

--document-version "\$LATEST"

--document-version "3"

$sel:documentHashType:MaintenanceWindowRunCommandParameters', maintenanceWindowRunCommandParameters_documentHashType - SHA-256 or SHA-1. SHA-1 hashes have been deprecated.

$sel:parameters:MaintenanceWindowRunCommandParameters', maintenanceWindowRunCommandParameters_parameters - The parameters for the RUN_COMMAND task execution.

maintenanceWindowRunCommandParameters_notificationConfig :: Lens' MaintenanceWindowRunCommandParameters (Maybe NotificationConfig)

Configurations for sending notifications about command status changes on a per-instance basis.

maintenanceWindowRunCommandParameters_serviceRoleArn :: Lens' MaintenanceWindowRunCommandParameters (Maybe Text)

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) service role to use to publish Amazon Simple Notification Service (Amazon SNS) notifications for maintenance window Run Command tasks.

maintenanceWindowRunCommandParameters_outputS3BucketName :: Lens' MaintenanceWindowRunCommandParameters (Maybe Text)

The name of the Amazon Simple Storage Service (Amazon S3) bucket.

maintenanceWindowRunCommandParameters_comment :: Lens' MaintenanceWindowRunCommandParameters (Maybe Text)

Information about the commands to run.

maintenanceWindowRunCommandParameters_documentHash :: Lens' MaintenanceWindowRunCommandParameters (Maybe Text)

The SHA-256 or SHA-1 hash created by the system when the document was created. SHA-1 hashes have been deprecated.

maintenanceWindowRunCommandParameters_timeoutSeconds :: Lens' MaintenanceWindowRunCommandParameters (Maybe Natural)

If this time is reached and the command hasn't already started running, it doesn't run.

maintenanceWindowRunCommandParameters_outputS3KeyPrefix :: Lens' MaintenanceWindowRunCommandParameters (Maybe Text)

The S3 bucket subfolder.

maintenanceWindowRunCommandParameters_cloudWatchOutputConfig :: Lens' MaintenanceWindowRunCommandParameters (Maybe CloudWatchOutputConfig)

Undocumented member.

maintenanceWindowRunCommandParameters_documentVersion :: Lens' MaintenanceWindowRunCommandParameters (Maybe Text)

The Amazon Web Services Systems Manager document (SSM document) version to use in the request. You can specify $DEFAULT, $LATEST, or a specific version number. If you run commands by using the Amazon Web Services CLI, then you must escape the first two options by using a backslash. If you specify a version number, then you don't need to use the backslash. For example:

--document-version "\$DEFAULT"

--document-version "\$LATEST"

--document-version "3"

maintenanceWindowRunCommandParameters_documentHashType :: Lens' MaintenanceWindowRunCommandParameters (Maybe DocumentHashType)

SHA-256 or SHA-1. SHA-1 hashes have been deprecated.

maintenanceWindowRunCommandParameters_parameters :: Lens' MaintenanceWindowRunCommandParameters (Maybe (HashMap Text [Text]))

The parameters for the RUN_COMMAND task execution.