Network.AWS.SSM.Types.MaintenanceWindowTarget

Description

Documentation

data MaintenanceWindowTarget

The target registered with the maintenance window.

See: newMaintenanceWindowTarget smart constructor.

Constructors

MaintenanceWindowTarget'

Fields

windowTargetId :: Maybe Text

The ID of the target.

resourceType :: Maybe MaintenanceWindowResourceType

The type of target that is being registered with the maintenance window.

targets :: Maybe [Target]

The targets, either instances or tags.

Specify instances using the following format:

Key=instanceids,Values=<instanceid1>,<instanceid2>

Tags are specified using the following format:

Key=<tag name>,Values=<tag value>.

name :: Maybe Text

The name for the maintenance window target.

windowId :: Maybe Text

The ID of the maintenance window to register the target with.

description :: Maybe (Sensitive Text)

A description for the target.

ownerInformation :: Maybe (Sensitive Text)

A user-provided value that will be included in any Amazon CloudWatch Events events that are raised while running tasks for these targets in this maintenance window.

newMaintenanceWindowTarget :: MaintenanceWindowTarget

Create a value of MaintenanceWindowTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:windowTargetId:MaintenanceWindowTarget', maintenanceWindowTarget_windowTargetId - The ID of the target.

$sel:resourceType:MaintenanceWindowTarget', maintenanceWindowTarget_resourceType - The type of target that is being registered with the maintenance window.

$sel:targets:MaintenanceWindowTarget', maintenanceWindowTarget_targets - The targets, either instances or tags.

Specify instances using the following format:

Key=instanceids,Values=<instanceid1>,<instanceid2>

Tags are specified using the following format:

Key=<tag name>,Values=<tag value>.

$sel:name:MaintenanceWindowTarget', maintenanceWindowTarget_name - The name for the maintenance window target.

$sel:windowId:MaintenanceWindowTarget', maintenanceWindowTarget_windowId - The ID of the maintenance window to register the target with.

$sel:description:MaintenanceWindowTarget', maintenanceWindowTarget_description - A description for the target.

$sel:ownerInformation:MaintenanceWindowTarget', maintenanceWindowTarget_ownerInformation - A user-provided value that will be included in any Amazon CloudWatch Events events that are raised while running tasks for these targets in this maintenance window.

maintenanceWindowTarget_windowTargetId :: Lens' MaintenanceWindowTarget (Maybe Text)

The ID of the target.

maintenanceWindowTarget_resourceType :: Lens' MaintenanceWindowTarget (Maybe MaintenanceWindowResourceType)

The type of target that is being registered with the maintenance window.

maintenanceWindowTarget_targets :: Lens' MaintenanceWindowTarget (Maybe [Target])

The targets, either instances or tags.

Specify instances using the following format:

Key=instanceids,Values=<instanceid1>,<instanceid2>

Tags are specified using the following format:

Key=<tag name>,Values=<tag value>.

maintenanceWindowTarget_name :: Lens' MaintenanceWindowTarget (Maybe Text)

The name for the maintenance window target.

maintenanceWindowTarget_windowId :: Lens' MaintenanceWindowTarget (Maybe Text)

The ID of the maintenance window to register the target with.

maintenanceWindowTarget_description :: Lens' MaintenanceWindowTarget (Maybe Text)

A description for the target.

maintenanceWindowTarget_ownerInformation :: Lens' MaintenanceWindowTarget (Maybe Text)

A user-provided value that will be included in any Amazon CloudWatch Events events that are raised while running tasks for these targets in this maintenance window.