Network.AWS.SSM.Types.MaintenanceWindowTask

Description

Documentation

data MaintenanceWindowTask

Information about a task defined for a maintenance window.

See: newMaintenanceWindowTask smart constructor.

Constructors

MaintenanceWindowTask'

Fields

maxErrors :: Maybe Text

The maximum number of errors allowed before this task stops being scheduled.

taskParameters :: Maybe (Sensitive (HashMap Text (Sensitive MaintenanceWindowTaskParameterValueExpression)))

The parameters that should be passed to the task when it is run.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

serviceRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) service role to use to publish Amazon Simple Notification Service (Amazon SNS) notifications for maintenance window Run Command tasks.

windowTaskId :: Maybe Text

The task ID.

cutoffBehavior :: Maybe MaintenanceWindowTaskCutoffBehavior

The specification for whether tasks should continue to run after the cutoff time specified in the maintenance windows is reached.

targets :: Maybe [Target]

The targets (either instances or tags). Instances are specified using Key=instanceids,Values=<instanceid1>,<instanceid2>. Tags are specified using Key=<tag name>,Values=<tag value>.

priority :: Maybe Natural

The priority of the task in the maintenance window. The lower the number, the higher the priority. Tasks that have the same priority are scheduled in parallel.

name :: Maybe Text

The task name.

windowId :: Maybe Text

The ID of the maintenance window where the task is registered.

maxConcurrency :: Maybe Text

The maximum number of targets this task can be run for, in parallel.

description :: Maybe (Sensitive Text)

A description of the task.

taskArn :: Maybe Text

The resource that the task uses during execution. For RUN_COMMAND and AUTOMATION task types, TaskArn is the Amazon Web Services Systems Manager (SSM document) name or ARN. For LAMBDA tasks, it's the function name or ARN. For STEP_FUNCTIONS tasks, it's the state machine ARN.

type' :: Maybe MaintenanceWindowTaskType

The type of task.

loggingInfo :: Maybe LoggingInfo

Information about an S3 bucket to write task-level logs to.

LoggingInfo has been deprecated. To specify an Amazon Simple Storage Service (Amazon S3) bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Amazon Web Services Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

newMaintenanceWindowTask :: MaintenanceWindowTask

Create a value of MaintenanceWindowTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxErrors:MaintenanceWindowTask', maintenanceWindowTask_maxErrors - The maximum number of errors allowed before this task stops being scheduled.

$sel:taskParameters:MaintenanceWindowTask', maintenanceWindowTask_taskParameters - The parameters that should be passed to the task when it is run.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

$sel:serviceRoleArn:MaintenanceWindowTask', maintenanceWindowTask_serviceRoleArn - The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) service role to use to publish Amazon Simple Notification Service (Amazon SNS) notifications for maintenance window Run Command tasks.

$sel:windowTaskId:MaintenanceWindowTask', maintenanceWindowTask_windowTaskId - The task ID.

$sel:cutoffBehavior:MaintenanceWindowTask', maintenanceWindowTask_cutoffBehavior - The specification for whether tasks should continue to run after the cutoff time specified in the maintenance windows is reached.

$sel:targets:MaintenanceWindowTask', maintenanceWindowTask_targets - The targets (either instances or tags). Instances are specified using Key=instanceids,Values=<instanceid1>,<instanceid2>. Tags are specified using Key=<tag name>,Values=<tag value>.

$sel:priority:MaintenanceWindowTask', maintenanceWindowTask_priority - The priority of the task in the maintenance window. The lower the number, the higher the priority. Tasks that have the same priority are scheduled in parallel.

$sel:name:MaintenanceWindowTask', maintenanceWindowTask_name - The task name.

$sel:windowId:MaintenanceWindowTask', maintenanceWindowTask_windowId - The ID of the maintenance window where the task is registered.

$sel:maxConcurrency:MaintenanceWindowTask', maintenanceWindowTask_maxConcurrency - The maximum number of targets this task can be run for, in parallel.

$sel:description:MaintenanceWindowTask', maintenanceWindowTask_description - A description of the task.

$sel:taskArn:MaintenanceWindowTask', maintenanceWindowTask_taskArn - The resource that the task uses during execution. For RUN_COMMAND and AUTOMATION task types, TaskArn is the Amazon Web Services Systems Manager (SSM document) name or ARN. For LAMBDA tasks, it's the function name or ARN. For STEP_FUNCTIONS tasks, it's the state machine ARN.

$sel:type':MaintenanceWindowTask', maintenanceWindowTask_type - The type of task.

$sel:loggingInfo:MaintenanceWindowTask', maintenanceWindowTask_loggingInfo - Information about an S3 bucket to write task-level logs to.

LoggingInfo has been deprecated. To specify an Amazon Simple Storage Service (Amazon S3) bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Amazon Web Services Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

maintenanceWindowTask_maxErrors :: Lens' MaintenanceWindowTask (Maybe Text)

The maximum number of errors allowed before this task stops being scheduled.

maintenanceWindowTask_taskParameters :: Lens' MaintenanceWindowTask (Maybe (HashMap Text MaintenanceWindowTaskParameterValueExpression))

The parameters that should be passed to the task when it is run.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

maintenanceWindowTask_serviceRoleArn :: Lens' MaintenanceWindowTask (Maybe Text)

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) service role to use to publish Amazon Simple Notification Service (Amazon SNS) notifications for maintenance window Run Command tasks.

maintenanceWindowTask_windowTaskId :: Lens' MaintenanceWindowTask (Maybe Text)

The task ID.

maintenanceWindowTask_cutoffBehavior :: Lens' MaintenanceWindowTask (Maybe MaintenanceWindowTaskCutoffBehavior)

The specification for whether tasks should continue to run after the cutoff time specified in the maintenance windows is reached.

maintenanceWindowTask_targets :: Lens' MaintenanceWindowTask (Maybe [Target])

The targets (either instances or tags). Instances are specified using Key=instanceids,Values=<instanceid1>,<instanceid2>. Tags are specified using Key=<tag name>,Values=<tag value>.

maintenanceWindowTask_priority :: Lens' MaintenanceWindowTask (Maybe Natural)

The priority of the task in the maintenance window. The lower the number, the higher the priority. Tasks that have the same priority are scheduled in parallel.

maintenanceWindowTask_name :: Lens' MaintenanceWindowTask (Maybe Text)

The task name.

maintenanceWindowTask_windowId :: Lens' MaintenanceWindowTask (Maybe Text)

The ID of the maintenance window where the task is registered.

maintenanceWindowTask_maxConcurrency :: Lens' MaintenanceWindowTask (Maybe Text)

The maximum number of targets this task can be run for, in parallel.

maintenanceWindowTask_description :: Lens' MaintenanceWindowTask (Maybe Text)

A description of the task.

maintenanceWindowTask_taskArn :: Lens' MaintenanceWindowTask (Maybe Text)

The resource that the task uses during execution. For RUN_COMMAND and AUTOMATION task types, TaskArn is the Amazon Web Services Systems Manager (SSM document) name or ARN. For LAMBDA tasks, it's the function name or ARN. For STEP_FUNCTIONS tasks, it's the state machine ARN.

maintenanceWindowTask_type :: Lens' MaintenanceWindowTask (Maybe MaintenanceWindowTaskType)

The type of task.

maintenanceWindowTask_loggingInfo :: Lens' MaintenanceWindowTask (Maybe LoggingInfo)

Information about an S3 bucket to write task-level logs to.

LoggingInfo has been deprecated. To specify an Amazon Simple Storage Service (Amazon S3) bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Amazon Web Services Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.