Network.AWS.SSM.Types.NotificationConfig

Description

Documentation

data NotificationConfig

Configurations for sending notifications.

See: newNotificationConfig smart constructor.

Constructors

NotificationConfig'

Fields

notificationEvents :: Maybe [NotificationEvent]

The different events for which you can receive notifications. To learn more about these events, see Monitoring Systems Manager status changes using Amazon SNS notifications in the Amazon Web Services Systems Manager User Guide.

notificationType :: Maybe NotificationType

The type of notification.

Command: Receive notification when the status of a command changes.

Invocation: For commands sent to multiple instances, receive notification on a per-instance basis when the status of a command changes.

notificationArn :: Maybe Text

An Amazon Resource Name (ARN) for an Amazon Simple Notification Service (Amazon SNS) topic. Run Command pushes notifications about command status changes to this topic.

newNotificationConfig :: NotificationConfig

Create a value of NotificationConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notificationEvents:NotificationConfig', notificationConfig_notificationEvents - The different events for which you can receive notifications. To learn more about these events, see Monitoring Systems Manager status changes using Amazon SNS notifications in the Amazon Web Services Systems Manager User Guide.

$sel:notificationType:NotificationConfig', notificationConfig_notificationType - The type of notification.

Command: Receive notification when the status of a command changes.

Invocation: For commands sent to multiple instances, receive notification on a per-instance basis when the status of a command changes.

$sel:notificationArn:NotificationConfig', notificationConfig_notificationArn - An Amazon Resource Name (ARN) for an Amazon Simple Notification Service (Amazon SNS) topic. Run Command pushes notifications about command status changes to this topic.

notificationConfig_notificationEvents :: Lens' NotificationConfig (Maybe [NotificationEvent])

The different events for which you can receive notifications. To learn more about these events, see Monitoring Systems Manager status changes using Amazon SNS notifications in the Amazon Web Services Systems Manager User Guide.

notificationConfig_notificationType :: Lens' NotificationConfig (Maybe NotificationType)

The type of notification.

Command: Receive notification when the status of a command changes.

Invocation: For commands sent to multiple instances, receive notification on a per-instance basis when the status of a command changes.

notificationConfig_notificationArn :: Lens' NotificationConfig (Maybe Text)

An Amazon Resource Name (ARN) for an Amazon Simple Notification Service (Amazon SNS) topic. Run Command pushes notifications about command status changes to this topic.