Network.AWS.SSM.Types.OpsAggregator

Description

Documentation

data OpsAggregator

One or more aggregators for viewing counts of OpsData using different dimensions such as Source, CreatedTime, or Source and CreatedTime, to name a few.

See: newOpsAggregator smart constructor.

Constructors

OpsAggregator'

Fields

typeName :: Maybe Text

The data type name to use for viewing counts of OpsData.

aggregators :: Maybe (NonEmpty OpsAggregator)

A nested aggregator for viewing counts of OpsData.

values :: Maybe (HashMap Text Text)

The aggregator value.

filters :: Maybe (NonEmpty OpsFilter)

The aggregator filters.

attributeName :: Maybe Text

The name of an OpsData attribute on which to limit the count of OpsData.

aggregatorType :: Maybe Text

Either a Range or Count aggregator for limiting an OpsData summary.

newOpsAggregator :: OpsAggregator

Create a value of OpsAggregator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeName:OpsAggregator', opsAggregator_typeName - The data type name to use for viewing counts of OpsData.

$sel:aggregators:OpsAggregator', opsAggregator_aggregators - A nested aggregator for viewing counts of OpsData.

$sel:values:OpsAggregator', opsAggregator_values - The aggregator value.

$sel:filters:OpsAggregator', opsAggregator_filters - The aggregator filters.

$sel:attributeName:OpsAggregator', opsAggregator_attributeName - The name of an OpsData attribute on which to limit the count of OpsData.

$sel:aggregatorType:OpsAggregator', opsAggregator_aggregatorType - Either a Range or Count aggregator for limiting an OpsData summary.

opsAggregator_typeName :: Lens' OpsAggregator (Maybe Text)

The data type name to use for viewing counts of OpsData.

opsAggregator_aggregators :: Lens' OpsAggregator (Maybe (NonEmpty OpsAggregator))

A nested aggregator for viewing counts of OpsData.

opsAggregator_values :: Lens' OpsAggregator (Maybe (HashMap Text Text))

The aggregator value.

opsAggregator_filters :: Lens' OpsAggregator (Maybe (NonEmpty OpsFilter))

The aggregator filters.

opsAggregator_attributeName :: Lens' OpsAggregator (Maybe Text)

The name of an OpsData attribute on which to limit the count of OpsData.

opsAggregator_aggregatorType :: Lens' OpsAggregator (Maybe Text)

Either a Range or Count aggregator for limiting an OpsData summary.