Network.AWS.SSM.Types.OpsItem

Description

Documentation

data OpsItem

Operations engineers and IT professionals use Amazon Web Services Systems Manager OpsCenter to view, investigate, and remediate operational work items (OpsItems) impacting the performance and health of their Amazon Web Services resources. OpsCenter is integrated with Amazon EventBridge and Amazon CloudWatch. This means you can configure these services to automatically create an OpsItem in OpsCenter when a CloudWatch alarm enters the ALARM state or when EventBridge processes an event from any Amazon Web Services service that publishes events. Configuring Amazon CloudWatch alarms and EventBridge events to automatically create OpsItems allows you to quickly diagnose and remediate issues with Amazon Web Services resources from a single console.

To help you diagnose issues, each OpsItem includes contextually relevant information such as the name and ID of the Amazon Web Services resource that generated the OpsItem, alarm or event details, alarm history, and an alarm timeline graph. For the Amazon Web Services resource, OpsCenter aggregates information from Config, CloudTrail logs, and EventBridge, so you don't have to navigate across multiple console pages during your investigation. For more information, see OpsCenter in the Amazon Web Services Systems Manager User Guide.

See: newOpsItem smart constructor.

Constructors

OpsItem'

Fields

status :: Maybe OpsItemStatus

The OpsItem status. Status can be Open, In Progress, or Resolved. For more information, see Editing OpsItem details in the Amazon Web Services Systems Manager User Guide.

plannedEndTime :: Maybe POSIX

The time specified in a change request for a runbook workflow to end. Currently supported only for the OpsItem type /aws/changerequest.

severity :: Maybe Text

The severity of the OpsItem. Severity options range from 1 to 4.

actualStartTime :: Maybe POSIX

The time a runbook workflow started. Currently reported only for the OpsItem type /aws/changerequest.

category :: Maybe Text

An OpsItem category. Category options include: Availability, Cost, Performance, Recovery, Security.

operationalData :: Maybe (HashMap Text OpsItemDataValue)

Operational data is custom data that provides useful reference details about the OpsItem. For example, you can specify log files, error strings, license keys, troubleshooting tips, or other relevant data. You enter operational data as key-value pairs. The key has a maximum length of 128 characters. The value has a maximum size of 20 KB.

Operational data keys can't begin with the following: amazon, aws, amzn, ssm, /amazon, /aws, /amzn, /ssm.

You can choose to make the data searchable by other users in the account or you can restrict search access. Searchable data means that all users with access to the OpsItem Overview page (as provided by the DescribeOpsItems API operation) can view and search on the specified data. Operational data that isn't searchable is only viewable by users who have access to the OpsItem (as provided by the GetOpsItem API operation).

Use the /aws/resources key in OperationalData to specify a related resource in the request. Use the /aws/automations key in OperationalData to associate an Automation runbook with the OpsItem. To view Amazon Web Services CLI example commands that use these keys, see Creating OpsItems manually in the Amazon Web Services Systems Manager User Guide.

title :: Maybe Text

A short heading that describes the nature of the OpsItem and the impacted resource.

source :: Maybe Text

The origin of the OpsItem, such as Amazon EC2 or Systems Manager. The impacted resource is a subset of source.

createdTime :: Maybe POSIX

The date and time the OpsItem was created.

version :: Maybe Text

The version of this OpsItem. Each time the OpsItem is edited the version number increments by one.

priority :: Maybe Natural

The importance of this OpsItem in relation to other OpsItems in the system.

actualEndTime :: Maybe POSIX

The time a runbook workflow ended. Currently reported only for the OpsItem type /aws/changerequest.

opsItemId :: Maybe Text

The ID of the OpsItem.

plannedStartTime :: Maybe POSIX

The time specified in a change request for a runbook workflow to start. Currently supported only for the OpsItem type /aws/changerequest.

opsItemType :: Maybe Text

The type of OpsItem. Currently, the only valid values are /aws/changerequest and /aws/issue.

notifications :: Maybe [OpsItemNotification]

The Amazon Resource Name (ARN) of an Amazon Simple Notification Service (Amazon SNS) topic where notifications are sent when this OpsItem is edited or changed.

lastModifiedTime :: Maybe POSIX

The date and time the OpsItem was last updated.

description :: Maybe Text

The OpsItem description.

createdBy :: Maybe Text

The ARN of the Amazon Web Services account that created the OpsItem.

lastModifiedBy :: Maybe Text

The ARN of the Amazon Web Services account that last updated the OpsItem.

relatedOpsItems :: Maybe [RelatedOpsItem]

One or more OpsItems that share something in common with the current OpsItem. For example, related OpsItems can include OpsItems with similar error messages, impacted resources, or statuses for the impacted resource.

newOpsItem :: OpsItem

Create a value of OpsItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:OpsItem', opsItem_status - The OpsItem status. Status can be Open, In Progress, or Resolved. For more information, see Editing OpsItem details in the Amazon Web Services Systems Manager User Guide.

$sel:plannedEndTime:OpsItem', opsItem_plannedEndTime - The time specified in a change request for a runbook workflow to end. Currently supported only for the OpsItem type /aws/changerequest.

$sel:severity:OpsItem', opsItem_severity - The severity of the OpsItem. Severity options range from 1 to 4.

$sel:actualStartTime:OpsItem', opsItem_actualStartTime - The time a runbook workflow started. Currently reported only for the OpsItem type /aws/changerequest.

$sel:category:OpsItem', opsItem_category - An OpsItem category. Category options include: Availability, Cost, Performance, Recovery, Security.

$sel:operationalData:OpsItem', opsItem_operationalData - Operational data is custom data that provides useful reference details about the OpsItem. For example, you can specify log files, error strings, license keys, troubleshooting tips, or other relevant data. You enter operational data as key-value pairs. The key has a maximum length of 128 characters. The value has a maximum size of 20 KB.

Operational data keys can't begin with the following: amazon, aws, amzn, ssm, /amazon, /aws, /amzn, /ssm.

You can choose to make the data searchable by other users in the account or you can restrict search access. Searchable data means that all users with access to the OpsItem Overview page (as provided by the DescribeOpsItems API operation) can view and search on the specified data. Operational data that isn't searchable is only viewable by users who have access to the OpsItem (as provided by the GetOpsItem API operation).

Use the /aws/resources key in OperationalData to specify a related resource in the request. Use the /aws/automations key in OperationalData to associate an Automation runbook with the OpsItem. To view Amazon Web Services CLI example commands that use these keys, see Creating OpsItems manually in the Amazon Web Services Systems Manager User Guide.

$sel:title:OpsItem', opsItem_title - A short heading that describes the nature of the OpsItem and the impacted resource.

$sel:source:OpsItem', opsItem_source - The origin of the OpsItem, such as Amazon EC2 or Systems Manager. The impacted resource is a subset of source.

$sel:createdTime:OpsItem', opsItem_createdTime - The date and time the OpsItem was created.

$sel:version:OpsItem', opsItem_version - The version of this OpsItem. Each time the OpsItem is edited the version number increments by one.

$sel:priority:OpsItem', opsItem_priority - The importance of this OpsItem in relation to other OpsItems in the system.

$sel:actualEndTime:OpsItem', opsItem_actualEndTime - The time a runbook workflow ended. Currently reported only for the OpsItem type /aws/changerequest.

$sel:opsItemId:OpsItem', opsItem_opsItemId - The ID of the OpsItem.

$sel:plannedStartTime:OpsItem', opsItem_plannedStartTime - The time specified in a change request for a runbook workflow to start. Currently supported only for the OpsItem type /aws/changerequest.

$sel:opsItemType:OpsItem', opsItem_opsItemType - The type of OpsItem. Currently, the only valid values are /aws/changerequest and /aws/issue.

$sel:notifications:OpsItem', opsItem_notifications - The Amazon Resource Name (ARN) of an Amazon Simple Notification Service (Amazon SNS) topic where notifications are sent when this OpsItem is edited or changed.

$sel:lastModifiedTime:OpsItem', opsItem_lastModifiedTime - The date and time the OpsItem was last updated.

$sel:description:OpsItem', opsItem_description - The OpsItem description.

$sel:createdBy:OpsItem', opsItem_createdBy - The ARN of the Amazon Web Services account that created the OpsItem.

$sel:lastModifiedBy:OpsItem', opsItem_lastModifiedBy - The ARN of the Amazon Web Services account that last updated the OpsItem.

$sel:relatedOpsItems:OpsItem', opsItem_relatedOpsItems - One or more OpsItems that share something in common with the current OpsItem. For example, related OpsItems can include OpsItems with similar error messages, impacted resources, or statuses for the impacted resource.

opsItem_status :: Lens' OpsItem (Maybe OpsItemStatus)

The OpsItem status. Status can be Open, In Progress, or Resolved. For more information, see Editing OpsItem details in the Amazon Web Services Systems Manager User Guide.

opsItem_plannedEndTime :: Lens' OpsItem (Maybe UTCTime)

The time specified in a change request for a runbook workflow to end. Currently supported only for the OpsItem type /aws/changerequest.

opsItem_severity :: Lens' OpsItem (Maybe Text)

The severity of the OpsItem. Severity options range from 1 to 4.

opsItem_actualStartTime :: Lens' OpsItem (Maybe UTCTime)

The time a runbook workflow started. Currently reported only for the OpsItem type /aws/changerequest.

opsItem_category :: Lens' OpsItem (Maybe Text)

An OpsItem category. Category options include: Availability, Cost, Performance, Recovery, Security.

opsItem_operationalData :: Lens' OpsItem (Maybe (HashMap Text OpsItemDataValue))

Operational data is custom data that provides useful reference details about the OpsItem. For example, you can specify log files, error strings, license keys, troubleshooting tips, or other relevant data. You enter operational data as key-value pairs. The key has a maximum length of 128 characters. The value has a maximum size of 20 KB.

Operational data keys can't begin with the following: amazon, aws, amzn, ssm, /amazon, /aws, /amzn, /ssm.

You can choose to make the data searchable by other users in the account or you can restrict search access. Searchable data means that all users with access to the OpsItem Overview page (as provided by the DescribeOpsItems API operation) can view and search on the specified data. Operational data that isn't searchable is only viewable by users who have access to the OpsItem (as provided by the GetOpsItem API operation).

Use the /aws/resources key in OperationalData to specify a related resource in the request. Use the /aws/automations key in OperationalData to associate an Automation runbook with the OpsItem. To view Amazon Web Services CLI example commands that use these keys, see Creating OpsItems manually in the Amazon Web Services Systems Manager User Guide.

opsItem_title :: Lens' OpsItem (Maybe Text)

A short heading that describes the nature of the OpsItem and the impacted resource.

opsItem_source :: Lens' OpsItem (Maybe Text)

The origin of the OpsItem, such as Amazon EC2 or Systems Manager. The impacted resource is a subset of source.

opsItem_createdTime :: Lens' OpsItem (Maybe UTCTime)

The date and time the OpsItem was created.

opsItem_version :: Lens' OpsItem (Maybe Text)

The version of this OpsItem. Each time the OpsItem is edited the version number increments by one.

opsItem_priority :: Lens' OpsItem (Maybe Natural)

The importance of this OpsItem in relation to other OpsItems in the system.

opsItem_actualEndTime :: Lens' OpsItem (Maybe UTCTime)

The time a runbook workflow ended. Currently reported only for the OpsItem type /aws/changerequest.

opsItem_opsItemId :: Lens' OpsItem (Maybe Text)

The ID of the OpsItem.

opsItem_plannedStartTime :: Lens' OpsItem (Maybe UTCTime)

The time specified in a change request for a runbook workflow to start. Currently supported only for the OpsItem type /aws/changerequest.

opsItem_opsItemType :: Lens' OpsItem (Maybe Text)

The type of OpsItem. Currently, the only valid values are /aws/changerequest and /aws/issue.

opsItem_notifications :: Lens' OpsItem (Maybe [OpsItemNotification])

The Amazon Resource Name (ARN) of an Amazon Simple Notification Service (Amazon SNS) topic where notifications are sent when this OpsItem is edited or changed.

opsItem_lastModifiedTime :: Lens' OpsItem (Maybe UTCTime)

The date and time the OpsItem was last updated.

opsItem_description :: Lens' OpsItem (Maybe Text)

The OpsItem description.

opsItem_createdBy :: Lens' OpsItem (Maybe Text)

The ARN of the Amazon Web Services account that created the OpsItem.

opsItem_lastModifiedBy :: Lens' OpsItem (Maybe Text)

The ARN of the Amazon Web Services account that last updated the OpsItem.

opsItem_relatedOpsItems :: Lens' OpsItem (Maybe [RelatedOpsItem])

One or more OpsItems that share something in common with the current OpsItem. For example, related OpsItems can include OpsItems with similar error messages, impacted resources, or statuses for the impacted resource.