Network.AWS.SSM.Types.OpsItemEventSummary

Description

Documentation

data OpsItemEventSummary

Summary information about an OpsItem event or that associated an OpsItem with a related item.

See: newOpsItemEventSummary smart constructor.

Constructors

OpsItemEventSummary'

Fields

opsItemId :: Maybe Text

The ID of the OpsItem.

createdTime :: Maybe POSIX

The date and time the OpsItem event was created.

createdBy :: Maybe OpsItemIdentity

Information about the user or resource that created the OpsItem event.

detailType :: Maybe Text

The type of information provided as a detail.

source :: Maybe Text

The source of the OpsItem event.

detail :: Maybe Text

Specific information about the OpsItem event.

eventId :: Maybe Text

The ID of the OpsItem event.

newOpsItemEventSummary :: OpsItemEventSummary

Create a value of OpsItemEventSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:opsItemId:OpsItemEventSummary', opsItemEventSummary_opsItemId - The ID of the OpsItem.

$sel:createdTime:OpsItemEventSummary', opsItemEventSummary_createdTime - The date and time the OpsItem event was created.

$sel:createdBy:OpsItemEventSummary', opsItemEventSummary_createdBy - Information about the user or resource that created the OpsItem event.

$sel:detailType:OpsItemEventSummary', opsItemEventSummary_detailType - The type of information provided as a detail.

$sel:source:OpsItemEventSummary', opsItemEventSummary_source - The source of the OpsItem event.

$sel:detail:OpsItemEventSummary', opsItemEventSummary_detail - Specific information about the OpsItem event.

$sel:eventId:OpsItemEventSummary', opsItemEventSummary_eventId - The ID of the OpsItem event.

opsItemEventSummary_opsItemId :: Lens' OpsItemEventSummary (Maybe Text)

The ID of the OpsItem.

opsItemEventSummary_createdTime :: Lens' OpsItemEventSummary (Maybe UTCTime)

The date and time the OpsItem event was created.

opsItemEventSummary_createdBy :: Lens' OpsItemEventSummary (Maybe OpsItemIdentity)

Information about the user or resource that created the OpsItem event.

opsItemEventSummary_detailType :: Lens' OpsItemEventSummary (Maybe Text)

The type of information provided as a detail.

opsItemEventSummary_source :: Lens' OpsItemEventSummary (Maybe Text)

The source of the OpsItem event.

opsItemEventSummary_detail :: Lens' OpsItemEventSummary (Maybe Text)

Specific information about the OpsItem event.

opsItemEventSummary_eventId :: Lens' OpsItemEventSummary (Maybe Text)

The ID of the OpsItem event.