Network.AWS.SSM.Types.OpsItemSummary

Description

Documentation

data OpsItemSummary

A count of OpsItems.

See: newOpsItemSummary smart constructor.

Constructors

OpsItemSummary'

Fields

actualEndTime :: Maybe POSIX

The time a runbook workflow ended. Currently reported only for the OpsItem type /aws/changerequest.

opsItemId :: Maybe Text

The ID of the OpsItem.

status :: Maybe OpsItemStatus

The OpsItem status. Status can be Open, In Progress, or Resolved.

priority :: Maybe Natural

The importance of this OpsItem in relation to other OpsItems in the system.

createdTime :: Maybe POSIX

The date and time the OpsItem was created.

category :: Maybe Text

A list of OpsItems by category.

severity :: Maybe Text

A list of OpsItems by severity.

createdBy :: Maybe Text

The Amazon Resource Name (ARN) of the IAM entity that created the OpsItem.

lastModifiedTime :: Maybe POSIX

The date and time the OpsItem was last updated.

opsItemType :: Maybe Text

The type of OpsItem. Currently, the only valid values are /aws/changerequest and /aws/issue.

source :: Maybe Text

The impacted Amazon Web Services resource.

title :: Maybe Text

A short heading that describes the nature of the OpsItem and the impacted resource.

lastModifiedBy :: Maybe Text

The Amazon Resource Name (ARN) of the IAM entity that created the OpsItem.

operationalData :: Maybe (HashMap Text OpsItemDataValue)

Operational data is custom data that provides useful reference details about the OpsItem.

actualStartTime :: Maybe POSIX

The time a runbook workflow started. Currently reported only for the OpsItem type /aws/changerequest.

plannedEndTime :: Maybe POSIX

The time specified in a change request for a runbook workflow to end. Currently supported only for the OpsItem type /aws/changerequest.

plannedStartTime :: Maybe POSIX

The time specified in a change request for a runbook workflow to start. Currently supported only for the OpsItem type /aws/changerequest.

newOpsItemSummary :: OpsItemSummary

Create a value of OpsItemSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actualEndTime:OpsItemSummary', opsItemSummary_actualEndTime - The time a runbook workflow ended. Currently reported only for the OpsItem type /aws/changerequest.

$sel:opsItemId:OpsItemSummary', opsItemSummary_opsItemId - The ID of the OpsItem.

$sel:status:OpsItemSummary', opsItemSummary_status - The OpsItem status. Status can be Open, In Progress, or Resolved.

$sel:priority:OpsItemSummary', opsItemSummary_priority - The importance of this OpsItem in relation to other OpsItems in the system.

$sel:createdTime:OpsItemSummary', opsItemSummary_createdTime - The date and time the OpsItem was created.

$sel:category:OpsItemSummary', opsItemSummary_category - A list of OpsItems by category.

$sel:severity:OpsItemSummary', opsItemSummary_severity - A list of OpsItems by severity.

$sel:createdBy:OpsItemSummary', opsItemSummary_createdBy - The Amazon Resource Name (ARN) of the IAM entity that created the OpsItem.

$sel:lastModifiedTime:OpsItemSummary', opsItemSummary_lastModifiedTime - The date and time the OpsItem was last updated.

$sel:opsItemType:OpsItemSummary', opsItemSummary_opsItemType - The type of OpsItem. Currently, the only valid values are /aws/changerequest and /aws/issue.

$sel:source:OpsItemSummary', opsItemSummary_source - The impacted Amazon Web Services resource.

$sel:title:OpsItemSummary', opsItemSummary_title - A short heading that describes the nature of the OpsItem and the impacted resource.

$sel:lastModifiedBy:OpsItemSummary', opsItemSummary_lastModifiedBy - The Amazon Resource Name (ARN) of the IAM entity that created the OpsItem.

$sel:operationalData:OpsItemSummary', opsItemSummary_operationalData - Operational data is custom data that provides useful reference details about the OpsItem.

$sel:actualStartTime:OpsItemSummary', opsItemSummary_actualStartTime - The time a runbook workflow started. Currently reported only for the OpsItem type /aws/changerequest.

$sel:plannedEndTime:OpsItemSummary', opsItemSummary_plannedEndTime - The time specified in a change request for a runbook workflow to end. Currently supported only for the OpsItem type /aws/changerequest.

$sel:plannedStartTime:OpsItemSummary', opsItemSummary_plannedStartTime - The time specified in a change request for a runbook workflow to start. Currently supported only for the OpsItem type /aws/changerequest.

opsItemSummary_actualEndTime :: Lens' OpsItemSummary (Maybe UTCTime)

The time a runbook workflow ended. Currently reported only for the OpsItem type /aws/changerequest.

opsItemSummary_opsItemId :: Lens' OpsItemSummary (Maybe Text)

The ID of the OpsItem.

opsItemSummary_status :: Lens' OpsItemSummary (Maybe OpsItemStatus)

The OpsItem status. Status can be Open, In Progress, or Resolved.

opsItemSummary_priority :: Lens' OpsItemSummary (Maybe Natural)

The importance of this OpsItem in relation to other OpsItems in the system.

opsItemSummary_createdTime :: Lens' OpsItemSummary (Maybe UTCTime)

The date and time the OpsItem was created.

opsItemSummary_category :: Lens' OpsItemSummary (Maybe Text)

A list of OpsItems by category.

opsItemSummary_severity :: Lens' OpsItemSummary (Maybe Text)

A list of OpsItems by severity.

opsItemSummary_createdBy :: Lens' OpsItemSummary (Maybe Text)

The Amazon Resource Name (ARN) of the IAM entity that created the OpsItem.

opsItemSummary_lastModifiedTime :: Lens' OpsItemSummary (Maybe UTCTime)

The date and time the OpsItem was last updated.

opsItemSummary_opsItemType :: Lens' OpsItemSummary (Maybe Text)

The type of OpsItem. Currently, the only valid values are /aws/changerequest and /aws/issue.

opsItemSummary_source :: Lens' OpsItemSummary (Maybe Text)

The impacted Amazon Web Services resource.

opsItemSummary_title :: Lens' OpsItemSummary (Maybe Text)

A short heading that describes the nature of the OpsItem and the impacted resource.

opsItemSummary_lastModifiedBy :: Lens' OpsItemSummary (Maybe Text)

The Amazon Resource Name (ARN) of the IAM entity that created the OpsItem.

opsItemSummary_operationalData :: Lens' OpsItemSummary (Maybe (HashMap Text OpsItemDataValue))

Operational data is custom data that provides useful reference details about the OpsItem.

opsItemSummary_actualStartTime :: Lens' OpsItemSummary (Maybe UTCTime)

The time a runbook workflow started. Currently reported only for the OpsItem type /aws/changerequest.

opsItemSummary_plannedEndTime :: Lens' OpsItemSummary (Maybe UTCTime)

The time specified in a change request for a runbook workflow to end. Currently supported only for the OpsItem type /aws/changerequest.

opsItemSummary_plannedStartTime :: Lens' OpsItemSummary (Maybe UTCTime)

The time specified in a change request for a runbook workflow to start. Currently supported only for the OpsItem type /aws/changerequest.