Network.AWS.SSM.Types.Parameter

Description

Documentation

data Parameter

An Amazon Web Services Systems Manager parameter in Parameter Store.

See: newParameter smart constructor.

Constructors

Parameter'

Fields

lastModifiedDate :: Maybe POSIX

Date the parameter was last changed or updated and the parameter version was created.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the parameter.

version :: Maybe Integer

The parameter version.

name :: Maybe Text

The name of the parameter.

sourceResult :: Maybe Text

Applies to parameters that reference information in other Amazon Web Services services. SourceResult is the raw result or response from the source.

value :: Maybe (Sensitive Text)

The parameter value.

dataType :: Maybe Text

The data type of the parameter, such as text or aws:ec2:image. The default is text.

type' :: Maybe ParameterType

The type of parameter. Valid values include the following: String, StringList, and SecureString.

selector :: Maybe Text

Either the version number or the label used to retrieve the parameter value. Specify selectors by using one of the following formats:

parameter_name:version

parameter_name:label

newParameter :: Parameter

Create a value of Parameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:Parameter', parameter_lastModifiedDate - Date the parameter was last changed or updated and the parameter version was created.

$sel:arn:Parameter', parameter_arn - The Amazon Resource Name (ARN) of the parameter.

$sel:version:Parameter', parameter_version - The parameter version.

$sel:name:Parameter', parameter_name - The name of the parameter.

$sel:sourceResult:Parameter', parameter_sourceResult - Applies to parameters that reference information in other Amazon Web Services services. SourceResult is the raw result or response from the source.

$sel:value:Parameter', parameter_value - The parameter value.

$sel:dataType:Parameter', parameter_dataType - The data type of the parameter, such as text or aws:ec2:image. The default is text.

$sel:type':Parameter', parameter_type - The type of parameter. Valid values include the following: String, StringList, and SecureString.

$sel:selector:Parameter', parameter_selector - Either the version number or the label used to retrieve the parameter value. Specify selectors by using one of the following formats:

parameter_name:version

parameter_name:label

parameter_lastModifiedDate :: Lens' Parameter (Maybe UTCTime)

Date the parameter was last changed or updated and the parameter version was created.

parameter_arn :: Lens' Parameter (Maybe Text)

The Amazon Resource Name (ARN) of the parameter.

parameter_version :: Lens' Parameter (Maybe Integer)

The parameter version.

parameter_name :: Lens' Parameter (Maybe Text)

The name of the parameter.

parameter_sourceResult :: Lens' Parameter (Maybe Text)

Applies to parameters that reference information in other Amazon Web Services services. SourceResult is the raw result or response from the source.

parameter_value :: Lens' Parameter (Maybe Text)

The parameter value.

parameter_dataType :: Lens' Parameter (Maybe Text)

The data type of the parameter, such as text or aws:ec2:image. The default is text.

parameter_type :: Lens' Parameter (Maybe ParameterType)

The type of parameter. Valid values include the following: String, StringList, and SecureString.

parameter_selector :: Lens' Parameter (Maybe Text)

Either the version number or the label used to retrieve the parameter value. Specify selectors by using one of the following formats:

parameter_name:version

parameter_name:label