Network.AWS.SSM.Types.ParameterInlinePolicy

Description

Documentation

data ParameterInlinePolicy

One or more policies assigned to a parameter.

See: newParameterInlinePolicy smart constructor.

Constructors

ParameterInlinePolicy'

Fields

policyType :: Maybe Text

The type of policy. Parameter Store, a capablility of Amazon Web Services Systems Manager, supports the following policy types: Expiration, ExpirationNotification, and NoChangeNotification.

policyText :: Maybe Text

The JSON text of the policy.

policyStatus :: Maybe Text

The status of the policy. Policies report the following statuses: Pending (the policy hasn't been enforced or applied yet), Finished (the policy was applied), Failed (the policy wasn't applied), or InProgress (the policy is being applied now).

newParameterInlinePolicy :: ParameterInlinePolicy

Create a value of ParameterInlinePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyType:ParameterInlinePolicy', parameterInlinePolicy_policyType - The type of policy. Parameter Store, a capablility of Amazon Web Services Systems Manager, supports the following policy types: Expiration, ExpirationNotification, and NoChangeNotification.

$sel:policyText:ParameterInlinePolicy', parameterInlinePolicy_policyText - The JSON text of the policy.

$sel:policyStatus:ParameterInlinePolicy', parameterInlinePolicy_policyStatus - The status of the policy. Policies report the following statuses: Pending (the policy hasn't been enforced or applied yet), Finished (the policy was applied), Failed (the policy wasn't applied), or InProgress (the policy is being applied now).

parameterInlinePolicy_policyType :: Lens' ParameterInlinePolicy (Maybe Text)

The type of policy. Parameter Store, a capablility of Amazon Web Services Systems Manager, supports the following policy types: Expiration, ExpirationNotification, and NoChangeNotification.

parameterInlinePolicy_policyText :: Lens' ParameterInlinePolicy (Maybe Text)

The JSON text of the policy.

parameterInlinePolicy_policyStatus :: Lens' ParameterInlinePolicy (Maybe Text)

The status of the policy. Policies report the following statuses: Pending (the policy hasn't been enforced or applied yet), Finished (the policy was applied), Failed (the policy wasn't applied), or InProgress (the policy is being applied now).