Network.AWS.SSM.Types.ParameterMetadata

Description

Documentation

data ParameterMetadata

Metadata includes information like the ARN of the last user and the date/time the parameter was last used.

See: newParameterMetadata smart constructor.

Constructors

ParameterMetadata'

Fields

lastModifiedDate :: Maybe POSIX

Date the parameter was last changed or updated.

policies :: Maybe [ParameterInlinePolicy]

A list of policies associated with a parameter.

version :: Maybe Integer

The parameter version.

name :: Maybe Text

The parameter name.

description :: Maybe Text

Description of the parameter actions.

dataType :: Maybe Text

The data type of the parameter, such as text or aws:ec2:image. The default is text.

allowedPattern :: Maybe Text

A parameter name can include only the following letters and symbols.

a-zA-Z0-9_.-

type' :: Maybe ParameterType

The type of parameter. Valid parameter types include the following: String, StringList, and SecureString.

lastModifiedUser :: Maybe Text

Amazon Resource Name (ARN) of the Amazon Web Services user who last changed the parameter.

tier :: Maybe ParameterTier

The parameter tier.

keyId :: Maybe Text

The ID of the query key used for this parameter.

newParameterMetadata :: ParameterMetadata

Create a value of ParameterMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:ParameterMetadata', parameterMetadata_lastModifiedDate - Date the parameter was last changed or updated.

$sel:policies:ParameterMetadata', parameterMetadata_policies - A list of policies associated with a parameter.

$sel:version:ParameterMetadata', parameterMetadata_version - The parameter version.

$sel:name:ParameterMetadata', parameterMetadata_name - The parameter name.

$sel:description:ParameterMetadata', parameterMetadata_description - Description of the parameter actions.

$sel:dataType:ParameterMetadata', parameterMetadata_dataType - The data type of the parameter, such as text or aws:ec2:image. The default is text.

$sel:allowedPattern:ParameterMetadata', parameterMetadata_allowedPattern - A parameter name can include only the following letters and symbols.

a-zA-Z0-9_.-

$sel:type':ParameterMetadata', parameterMetadata_type - The type of parameter. Valid parameter types include the following: String, StringList, and SecureString.

$sel:lastModifiedUser:ParameterMetadata', parameterMetadata_lastModifiedUser - Amazon Resource Name (ARN) of the Amazon Web Services user who last changed the parameter.

$sel:tier:ParameterMetadata', parameterMetadata_tier - The parameter tier.

$sel:keyId:ParameterMetadata', parameterMetadata_keyId - The ID of the query key used for this parameter.

parameterMetadata_lastModifiedDate :: Lens' ParameterMetadata (Maybe UTCTime)

Date the parameter was last changed or updated.

parameterMetadata_policies :: Lens' ParameterMetadata (Maybe [ParameterInlinePolicy])

A list of policies associated with a parameter.

parameterMetadata_version :: Lens' ParameterMetadata (Maybe Integer)

The parameter version.

parameterMetadata_name :: Lens' ParameterMetadata (Maybe Text)

The parameter name.

parameterMetadata_description :: Lens' ParameterMetadata (Maybe Text)

Description of the parameter actions.

parameterMetadata_dataType :: Lens' ParameterMetadata (Maybe Text)

The data type of the parameter, such as text or aws:ec2:image. The default is text.

parameterMetadata_allowedPattern :: Lens' ParameterMetadata (Maybe Text)

A parameter name can include only the following letters and symbols.

a-zA-Z0-9_.-

parameterMetadata_type :: Lens' ParameterMetadata (Maybe ParameterType)

The type of parameter. Valid parameter types include the following: String, StringList, and SecureString.

parameterMetadata_lastModifiedUser :: Lens' ParameterMetadata (Maybe Text)

Amazon Resource Name (ARN) of the Amazon Web Services user who last changed the parameter.

parameterMetadata_tier :: Lens' ParameterMetadata (Maybe ParameterTier)

The parameter tier.

parameterMetadata_keyId :: Lens' ParameterMetadata (Maybe Text)

The ID of the query key used for this parameter.