Network.AWS.SSM.Types.Patch

Description

Documentation

data Patch

Represents metadata about a patch.

See: newPatch smart constructor.

Constructors

Patch'

Fields

msrcSeverity :: Maybe Text

The severity of the patch, such as Critical, Important, or Moderate. Applies to Windows patches only.

vendor :: Maybe Text

The name of the vendor providing the patch.

epoch :: Maybe Int

The epoch of the patch. For example in pkg-example-EE-20180914-2.2.amzn1.noarch, the epoch value is 20180914-2. Applies to Linux-based instances only.

product :: Maybe Text

The specific product the patch is applicable for. For example, WindowsServer2016 or AmazonLinux2018.03.

severity :: Maybe Text

The severity level of the patch. For example, CRITICAL or MODERATE.

title :: Maybe Text

The title of the patch.

id :: Maybe Text

The ID of the patch. Applies to Windows patches only.

This ID isn't the same as the Microsoft Knowledge Base ID.

productFamily :: Maybe Text

The product family the patch is applicable for. For example, Windows or Amazon Linux 2.

repository :: Maybe Text

The source patch repository for the operating system and version, such as trusty-security for Ubuntu Server 14.04 LTE and focal-security for Ubuntu Server 20.04 LTE. Applies to Linux-based instances only.

version :: Maybe Text

The version number of the patch. For example, in example-pkg-1.710.10-2.7.abcd.x86_64, the version number is indicated by -1. Applies to Linux-based instances only.

name :: Maybe Text

The name of the patch. Applies to Linux-based instances only.

bugzillaIds :: Maybe [Text]

The Bugzilla ID of the patch. For example, 1600646. Applies to Linux-based instances only.

msrcNumber :: Maybe Text

The ID of the Microsoft Security Response Center (MSRC) bulletin the patch is related to. For example, MS14-045. Applies to Windows patches only.

release :: Maybe Text

The particular release of a patch. For example, in pkg-example-EE-20180914-2.2.amzn1.noarch, the release is 2.amaz1. Applies to Linux-based instances only.

cVEIds :: Maybe [Text]

The Common Vulnerabilities and Exposures (CVE) ID of the patch. For example, CVE-2011-3192. Applies to Linux-based instances only.

description :: Maybe Text

The description of the patch.

classification :: Maybe Text

The classification of the patch. For example, SecurityUpdates, Updates, or CriticalUpdates.

advisoryIds :: Maybe [Text]

The Advisory ID of the patch. For example, RHSA-2020:3779. Applies to Linux-based instances only.

arch :: Maybe Text

The architecture of the patch. For example, in example-pkg-0.710.10-2.7.abcd.x86_64, the architecture is indicated by x86_64. Applies to Linux-based instances only.

releaseDate :: Maybe POSIX

The date the patch was released.

language :: Maybe Text

The language of the patch if it's language-specific.

kbNumber :: Maybe Text

The Microsoft Knowledge Base ID of the patch. Applies to Windows patches only.

contentUrl :: Maybe Text

The URL where more information can be obtained about the patch.

newPatch :: Patch

Create a value of Patch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:msrcSeverity:Patch', patch_msrcSeverity - The severity of the patch, such as Critical, Important, or Moderate. Applies to Windows patches only.

$sel:vendor:Patch', patch_vendor - The name of the vendor providing the patch.

$sel:epoch:Patch', patch_epoch - The epoch of the patch. For example in pkg-example-EE-20180914-2.2.amzn1.noarch, the epoch value is 20180914-2. Applies to Linux-based instances only.

$sel:product:Patch', patch_product - The specific product the patch is applicable for. For example, WindowsServer2016 or AmazonLinux2018.03.

$sel:severity:Patch', patch_severity - The severity level of the patch. For example, CRITICAL or MODERATE.

$sel:title:Patch', patch_title - The title of the patch.

$sel:id:Patch', patch_id - The ID of the patch. Applies to Windows patches only.

This ID isn't the same as the Microsoft Knowledge Base ID.

$sel:productFamily:Patch', patch_productFamily - The product family the patch is applicable for. For example, Windows or Amazon Linux 2.

$sel:repository:Patch', patch_repository - The source patch repository for the operating system and version, such as trusty-security for Ubuntu Server 14.04 LTE and focal-security for Ubuntu Server 20.04 LTE. Applies to Linux-based instances only.

$sel:version:Patch', patch_version - The version number of the patch. For example, in example-pkg-1.710.10-2.7.abcd.x86_64, the version number is indicated by -1. Applies to Linux-based instances only.

$sel:name:Patch', patch_name - The name of the patch. Applies to Linux-based instances only.

$sel:bugzillaIds:Patch', patch_bugzillaIds - The Bugzilla ID of the patch. For example, 1600646. Applies to Linux-based instances only.

$sel:msrcNumber:Patch', patch_msrcNumber - The ID of the Microsoft Security Response Center (MSRC) bulletin the patch is related to. For example, MS14-045. Applies to Windows patches only.

$sel:release:Patch', patch_release - The particular release of a patch. For example, in pkg-example-EE-20180914-2.2.amzn1.noarch, the release is 2.amaz1. Applies to Linux-based instances only.

$sel:cVEIds:Patch', patch_cVEIds - The Common Vulnerabilities and Exposures (CVE) ID of the patch. For example, CVE-2011-3192. Applies to Linux-based instances only.

$sel:description:Patch', patch_description - The description of the patch.

$sel:classification:Patch', patch_classification - The classification of the patch. For example, SecurityUpdates, Updates, or CriticalUpdates.

$sel:advisoryIds:Patch', patch_advisoryIds - The Advisory ID of the patch. For example, RHSA-2020:3779. Applies to Linux-based instances only.

$sel:arch:Patch', patch_arch - The architecture of the patch. For example, in example-pkg-0.710.10-2.7.abcd.x86_64, the architecture is indicated by x86_64. Applies to Linux-based instances only.

$sel:releaseDate:Patch', patch_releaseDate - The date the patch was released.

$sel:language:Patch', patch_language - The language of the patch if it's language-specific.

$sel:kbNumber:Patch', patch_kbNumber - The Microsoft Knowledge Base ID of the patch. Applies to Windows patches only.

$sel:contentUrl:Patch', patch_contentUrl - The URL where more information can be obtained about the patch.

patch_msrcSeverity :: Lens' Patch (Maybe Text)

The severity of the patch, such as Critical, Important, or Moderate. Applies to Windows patches only.

patch_vendor :: Lens' Patch (Maybe Text)

The name of the vendor providing the patch.

patch_epoch :: Lens' Patch (Maybe Int)

The epoch of the patch. For example in pkg-example-EE-20180914-2.2.amzn1.noarch, the epoch value is 20180914-2. Applies to Linux-based instances only.

patch_product :: Lens' Patch (Maybe Text)

The specific product the patch is applicable for. For example, WindowsServer2016 or AmazonLinux2018.03.

patch_severity :: Lens' Patch (Maybe Text)

The severity level of the patch. For example, CRITICAL or MODERATE.

patch_title :: Lens' Patch (Maybe Text)

The title of the patch.

patch_id :: Lens' Patch (Maybe Text)

The ID of the patch. Applies to Windows patches only.

This ID isn't the same as the Microsoft Knowledge Base ID.

patch_productFamily :: Lens' Patch (Maybe Text)

The product family the patch is applicable for. For example, Windows or Amazon Linux 2.

patch_repository :: Lens' Patch (Maybe Text)

The source patch repository for the operating system and version, such as trusty-security for Ubuntu Server 14.04 LTE and focal-security for Ubuntu Server 20.04 LTE. Applies to Linux-based instances only.

patch_version :: Lens' Patch (Maybe Text)

The version number of the patch. For example, in example-pkg-1.710.10-2.7.abcd.x86_64, the version number is indicated by -1. Applies to Linux-based instances only.

patch_name :: Lens' Patch (Maybe Text)

The name of the patch. Applies to Linux-based instances only.

patch_bugzillaIds :: Lens' Patch (Maybe [Text])

The Bugzilla ID of the patch. For example, 1600646. Applies to Linux-based instances only.

patch_msrcNumber :: Lens' Patch (Maybe Text)

The ID of the Microsoft Security Response Center (MSRC) bulletin the patch is related to. For example, MS14-045. Applies to Windows patches only.

patch_release :: Lens' Patch (Maybe Text)

The particular release of a patch. For example, in pkg-example-EE-20180914-2.2.amzn1.noarch, the release is 2.amaz1. Applies to Linux-based instances only.

patch_cVEIds :: Lens' Patch (Maybe [Text])

The Common Vulnerabilities and Exposures (CVE) ID of the patch. For example, CVE-2011-3192. Applies to Linux-based instances only.

patch_description :: Lens' Patch (Maybe Text)

The description of the patch.

patch_classification :: Lens' Patch (Maybe Text)

The classification of the patch. For example, SecurityUpdates, Updates, or CriticalUpdates.

patch_advisoryIds :: Lens' Patch (Maybe [Text])

The Advisory ID of the patch. For example, RHSA-2020:3779. Applies to Linux-based instances only.

patch_arch :: Lens' Patch (Maybe Text)

The architecture of the patch. For example, in example-pkg-0.710.10-2.7.abcd.x86_64, the architecture is indicated by x86_64. Applies to Linux-based instances only.

patch_releaseDate :: Lens' Patch (Maybe UTCTime)

The date the patch was released.

patch_language :: Lens' Patch (Maybe Text)

The language of the patch if it's language-specific.

patch_kbNumber :: Lens' Patch (Maybe Text)

The Microsoft Knowledge Base ID of the patch. Applies to Windows patches only.

patch_contentUrl :: Lens' Patch (Maybe Text)

The URL where more information can be obtained about the patch.