Network.AWS.SSM.Types.PatchBaselineIdentity

Description

Documentation

data PatchBaselineIdentity

Defines the basic information about a patch baseline.

See: newPatchBaselineIdentity smart constructor.

Constructors

PatchBaselineIdentity'

Fields

baselineName :: Maybe Text

The name of the patch baseline.

baselineId :: Maybe Text

The ID of the patch baseline.

defaultBaseline :: Maybe Bool

Whether this is the default baseline. Amazon Web Services Systems Manager supports creating multiple default patch baselines. For example, you can create a default patch baseline for each operating system.

baselineDescription :: Maybe Text

The description of the patch baseline.

operatingSystem :: Maybe OperatingSystem

Defines the operating system the patch baseline applies to. The default value is WINDOWS.

newPatchBaselineIdentity :: PatchBaselineIdentity

Create a value of PatchBaselineIdentity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baselineName:PatchBaselineIdentity', patchBaselineIdentity_baselineName - The name of the patch baseline.

$sel:baselineId:PatchBaselineIdentity', patchBaselineIdentity_baselineId - The ID of the patch baseline.

$sel:defaultBaseline:PatchBaselineIdentity', patchBaselineIdentity_defaultBaseline - Whether this is the default baseline. Amazon Web Services Systems Manager supports creating multiple default patch baselines. For example, you can create a default patch baseline for each operating system.

$sel:baselineDescription:PatchBaselineIdentity', patchBaselineIdentity_baselineDescription - The description of the patch baseline.

$sel:operatingSystem:PatchBaselineIdentity', patchBaselineIdentity_operatingSystem - Defines the operating system the patch baseline applies to. The default value is WINDOWS.

patchBaselineIdentity_baselineName :: Lens' PatchBaselineIdentity (Maybe Text)

The name of the patch baseline.

patchBaselineIdentity_baselineId :: Lens' PatchBaselineIdentity (Maybe Text)

The ID of the patch baseline.

patchBaselineIdentity_defaultBaseline :: Lens' PatchBaselineIdentity (Maybe Bool)

Whether this is the default baseline. Amazon Web Services Systems Manager supports creating multiple default patch baselines. For example, you can create a default patch baseline for each operating system.

patchBaselineIdentity_baselineDescription :: Lens' PatchBaselineIdentity (Maybe Text)

The description of the patch baseline.

patchBaselineIdentity_operatingSystem :: Lens' PatchBaselineIdentity (Maybe OperatingSystem)

Defines the operating system the patch baseline applies to. The default value is WINDOWS.