Network.AWS.SSM.Types.PatchComplianceData

Description

Documentation

data PatchComplianceData

Information about the state of a patch on a particular instance as it relates to the patch baseline used to patch the instance.

See: newPatchComplianceData smart constructor.

Constructors

PatchComplianceData'

Fields

cVEIds :: Maybe Text

The IDs of one or more Common Vulnerabilities and Exposure (CVE) issues that are resolved by the patch.

title :: Text

The title of the patch.

kBId :: Text

The operating system-specific ID of the patch.

classification :: Text

The classification of the patch, such as SecurityUpdates, Updates, and CriticalUpdates.

severity :: Text

The severity of the patchsuch as Critical, Important, and Moderate.

state :: PatchComplianceDataState

The state of the patch on the instance, such as INSTALLED or FAILED.

For descriptions of each patch state, see About patch compliance in the Amazon Web Services Systems Manager User Guide.

installedTime :: POSIX

The date/time the patch was installed on the instance. Not all operating systems provide this level of information.

newPatchComplianceData

Arguments

:: Text
$sel:title:PatchComplianceData'

-> Text
$sel:kBId:PatchComplianceData'

-> Text
$sel:classification:PatchComplianceData'

-> Text
$sel:severity:PatchComplianceData'

-> PatchComplianceDataState
$sel:state:PatchComplianceData'

-> UTCTime
$sel:installedTime:PatchComplianceData'

-> PatchComplianceData

Create a value of PatchComplianceData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cVEIds:PatchComplianceData', patchComplianceData_cVEIds - The IDs of one or more Common Vulnerabilities and Exposure (CVE) issues that are resolved by the patch.

$sel:title:PatchComplianceData', patchComplianceData_title - The title of the patch.

$sel:kBId:PatchComplianceData', patchComplianceData_kBId - The operating system-specific ID of the patch.

$sel:classification:PatchComplianceData', patchComplianceData_classification - The classification of the patch, such as SecurityUpdates, Updates, and CriticalUpdates.

$sel:severity:PatchComplianceData', patchComplianceData_severity - The severity of the patchsuch as Critical, Important, and Moderate.

$sel:state:PatchComplianceData', patchComplianceData_state - The state of the patch on the instance, such as INSTALLED or FAILED.

For descriptions of each patch state, see About patch compliance in the Amazon Web Services Systems Manager User Guide.

$sel:installedTime:PatchComplianceData', patchComplianceData_installedTime - The date/time the patch was installed on the instance. Not all operating systems provide this level of information.

patchComplianceData_cVEIds :: Lens' PatchComplianceData (Maybe Text)

The IDs of one or more Common Vulnerabilities and Exposure (CVE) issues that are resolved by the patch.

patchComplianceData_title :: Lens' PatchComplianceData Text

The title of the patch.

patchComplianceData_kBId :: Lens' PatchComplianceData Text

The operating system-specific ID of the patch.

patchComplianceData_classification :: Lens' PatchComplianceData Text

The classification of the patch, such as SecurityUpdates, Updates, and CriticalUpdates.

patchComplianceData_severity :: Lens' PatchComplianceData Text

The severity of the patchsuch as Critical, Important, and Moderate.

patchComplianceData_state :: Lens' PatchComplianceData PatchComplianceDataState

The state of the patch on the instance, such as INSTALLED or FAILED.

For descriptions of each patch state, see About patch compliance in the Amazon Web Services Systems Manager User Guide.

patchComplianceData_installedTime :: Lens' PatchComplianceData UTCTime

The date/time the patch was installed on the instance. Not all operating systems provide this level of information.