Network.AWS.SSM.Types.PatchGroupPatchBaselineMapping

Description

Documentation

data PatchGroupPatchBaselineMapping

The mapping between a patch group and the patch baseline the patch group is registered with.

See: newPatchGroupPatchBaselineMapping smart constructor.

Constructors

PatchGroupPatchBaselineMapping'

Fields

baselineIdentity :: Maybe PatchBaselineIdentity

The patch baseline the patch group is registered with.

patchGroup :: Maybe Text

The name of the patch group registered with the patch baseline.

newPatchGroupPatchBaselineMapping :: PatchGroupPatchBaselineMapping

Create a value of PatchGroupPatchBaselineMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baselineIdentity:PatchGroupPatchBaselineMapping', patchGroupPatchBaselineMapping_baselineIdentity - The patch baseline the patch group is registered with.

$sel:patchGroup:PatchGroupPatchBaselineMapping', patchGroupPatchBaselineMapping_patchGroup - The name of the patch group registered with the patch baseline.

patchGroupPatchBaselineMapping_baselineIdentity :: Lens' PatchGroupPatchBaselineMapping (Maybe PatchBaselineIdentity)

The patch baseline the patch group is registered with.

patchGroupPatchBaselineMapping_patchGroup :: Lens' PatchGroupPatchBaselineMapping (Maybe Text)

The name of the patch group registered with the patch baseline.