Network.AWS.SSM.Types.PatchStatus

Description

Documentation

data PatchStatus

Information about the approval status of a patch.

See: newPatchStatus smart constructor.

Constructors

PatchStatus'

Fields

approvalDate :: Maybe POSIX

The date the patch was approved (or will be approved if the status is PENDING_APPROVAL).

complianceLevel :: Maybe PatchComplianceLevel

The compliance severity level for a patch.

deploymentStatus :: Maybe PatchDeploymentStatus

The approval status of a patch.

newPatchStatus :: PatchStatus

Create a value of PatchStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:approvalDate:PatchStatus', patchStatus_approvalDate - The date the patch was approved (or will be approved if the status is PENDING_APPROVAL).

$sel:complianceLevel:PatchStatus', patchStatus_complianceLevel - The compliance severity level for a patch.

$sel:deploymentStatus:PatchStatus', patchStatus_deploymentStatus - The approval status of a patch.

patchStatus_approvalDate :: Lens' PatchStatus (Maybe UTCTime)

The date the patch was approved (or will be approved if the status is PENDING_APPROVAL).

patchStatus_complianceLevel :: Lens' PatchStatus (Maybe PatchComplianceLevel)

The compliance severity level for a patch.

patchStatus_deploymentStatus :: Lens' PatchStatus (Maybe PatchDeploymentStatus)

The approval status of a patch.