Network.AWS.SSM.Types.ProgressCounters

Description

Documentation

data ProgressCounters

An aggregate of step execution statuses displayed in the Amazon Web Services Systems Manager console for a multi-Region and multi-account Automation execution.

See: newProgressCounters smart constructor.

Constructors

ProgressCounters'

Fields

failedSteps :: Maybe Int

The total number of steps that failed to run in all specified Amazon Web Services Regions and Amazon Web Services accounts for the current Automation execution.

cancelledSteps :: Maybe Int

The total number of steps that the system cancelled in all specified Amazon Web Services Regions and Amazon Web Services accounts for the current Automation execution.

successSteps :: Maybe Int

The total number of steps that successfully completed in all specified Amazon Web Services Regions and Amazon Web Services accounts for the current Automation execution.

totalSteps :: Maybe Int

The total number of steps run in all specified Amazon Web Services Regions and Amazon Web Services accounts for the current Automation execution.

timedOutSteps :: Maybe Int

The total number of steps that timed out in all specified Amazon Web Services Regions and Amazon Web Services accounts for the current Automation execution.

newProgressCounters :: ProgressCounters

Create a value of ProgressCounters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedSteps:ProgressCounters', progressCounters_failedSteps - The total number of steps that failed to run in all specified Amazon Web Services Regions and Amazon Web Services accounts for the current Automation execution.

$sel:cancelledSteps:ProgressCounters', progressCounters_cancelledSteps - The total number of steps that the system cancelled in all specified Amazon Web Services Regions and Amazon Web Services accounts for the current Automation execution.

$sel:successSteps:ProgressCounters', progressCounters_successSteps - The total number of steps that successfully completed in all specified Amazon Web Services Regions and Amazon Web Services accounts for the current Automation execution.

$sel:totalSteps:ProgressCounters', progressCounters_totalSteps - The total number of steps run in all specified Amazon Web Services Regions and Amazon Web Services accounts for the current Automation execution.

$sel:timedOutSteps:ProgressCounters', progressCounters_timedOutSteps - The total number of steps that timed out in all specified Amazon Web Services Regions and Amazon Web Services accounts for the current Automation execution.

progressCounters_failedSteps :: Lens' ProgressCounters (Maybe Int)

The total number of steps that failed to run in all specified Amazon Web Services Regions and Amazon Web Services accounts for the current Automation execution.

progressCounters_cancelledSteps :: Lens' ProgressCounters (Maybe Int)

The total number of steps that the system cancelled in all specified Amazon Web Services Regions and Amazon Web Services accounts for the current Automation execution.

progressCounters_successSteps :: Lens' ProgressCounters (Maybe Int)

The total number of steps that successfully completed in all specified Amazon Web Services Regions and Amazon Web Services accounts for the current Automation execution.

progressCounters_totalSteps :: Lens' ProgressCounters (Maybe Int)

The total number of steps run in all specified Amazon Web Services Regions and Amazon Web Services accounts for the current Automation execution.

progressCounters_timedOutSteps :: Lens' ProgressCounters (Maybe Int)

The total number of steps that timed out in all specified Amazon Web Services Regions and Amazon Web Services accounts for the current Automation execution.