Network.AWS.SSM.Types.ResourceComplianceSummaryItem

Description

Documentation

data ResourceComplianceSummaryItem

Compliance summary information for a specific resource.

See: newResourceComplianceSummaryItem smart constructor.

Constructors

ResourceComplianceSummaryItem'

Fields

nonCompliantSummary :: Maybe NonCompliantSummary

A list of items that aren't compliant for the resource.

status :: Maybe ComplianceStatus

The compliance status for the resource.

resourceId :: Maybe Text

The resource ID.

resourceType :: Maybe Text

The resource type.

compliantSummary :: Maybe CompliantSummary

A list of items that are compliant for the resource.

executionSummary :: Maybe ComplianceExecutionSummary

Information about the execution.

overallSeverity :: Maybe ComplianceSeverity

The highest severity item found for the resource. The resource is compliant for this item.

complianceType :: Maybe Text

The compliance type.

newResourceComplianceSummaryItem :: ResourceComplianceSummaryItem

Create a value of ResourceComplianceSummaryItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nonCompliantSummary:ResourceComplianceSummaryItem', resourceComplianceSummaryItem_nonCompliantSummary - A list of items that aren't compliant for the resource.

$sel:status:ResourceComplianceSummaryItem', resourceComplianceSummaryItem_status - The compliance status for the resource.

$sel:resourceId:ResourceComplianceSummaryItem', resourceComplianceSummaryItem_resourceId - The resource ID.

$sel:resourceType:ResourceComplianceSummaryItem', resourceComplianceSummaryItem_resourceType - The resource type.

$sel:compliantSummary:ResourceComplianceSummaryItem', resourceComplianceSummaryItem_compliantSummary - A list of items that are compliant for the resource.

$sel:executionSummary:ResourceComplianceSummaryItem', resourceComplianceSummaryItem_executionSummary - Information about the execution.

$sel:overallSeverity:ResourceComplianceSummaryItem', resourceComplianceSummaryItem_overallSeverity - The highest severity item found for the resource. The resource is compliant for this item.

$sel:complianceType:ResourceComplianceSummaryItem', resourceComplianceSummaryItem_complianceType - The compliance type.

resourceComplianceSummaryItem_nonCompliantSummary :: Lens' ResourceComplianceSummaryItem (Maybe NonCompliantSummary)

A list of items that aren't compliant for the resource.

resourceComplianceSummaryItem_status :: Lens' ResourceComplianceSummaryItem (Maybe ComplianceStatus)

The compliance status for the resource.

resourceComplianceSummaryItem_resourceId :: Lens' ResourceComplianceSummaryItem (Maybe Text)

The resource ID.

resourceComplianceSummaryItem_resourceType :: Lens' ResourceComplianceSummaryItem (Maybe Text)

The resource type.

resourceComplianceSummaryItem_compliantSummary :: Lens' ResourceComplianceSummaryItem (Maybe CompliantSummary)

A list of items that are compliant for the resource.

resourceComplianceSummaryItem_executionSummary :: Lens' ResourceComplianceSummaryItem (Maybe ComplianceExecutionSummary)

Information about the execution.

resourceComplianceSummaryItem_overallSeverity :: Lens' ResourceComplianceSummaryItem (Maybe ComplianceSeverity)

The highest severity item found for the resource. The resource is compliant for this item.

resourceComplianceSummaryItem_complianceType :: Lens' ResourceComplianceSummaryItem (Maybe Text)

The compliance type.