Network.AWS.SSM.Types.ResourceDataSyncDestinationDataSharing

Description

Documentation

data ResourceDataSyncDestinationDataSharing

Synchronize Amazon Web Services Systems Manager Inventory data from multiple Amazon Web Services accounts defined in Organizations to a centralized Amazon S3 bucket. Data is synchronized to individual key prefixes in the central bucket. Each key prefix represents a different Amazon Web Services account ID.

See: newResourceDataSyncDestinationDataSharing smart constructor.

Constructors

ResourceDataSyncDestinationDataSharing'

Fields

destinationDataSharingType :: Maybe Text

The sharing data type. Only Organization is supported.

newResourceDataSyncDestinationDataSharing :: ResourceDataSyncDestinationDataSharing

Create a value of ResourceDataSyncDestinationDataSharing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationDataSharingType:ResourceDataSyncDestinationDataSharing', resourceDataSyncDestinationDataSharing_destinationDataSharingType - The sharing data type. Only Organization is supported.

resourceDataSyncDestinationDataSharing_destinationDataSharingType :: Lens' ResourceDataSyncDestinationDataSharing (Maybe Text)

The sharing data type. Only Organization is supported.