Network.AWS.SSM.Types.ResourceDataSyncItem

Description

Documentation

data ResourceDataSyncItem

Information about a resource data sync configuration, including its current status and last successful sync.

See: newResourceDataSyncItem smart constructor.

Constructors

ResourceDataSyncItem'

Fields

syncType :: Maybe Text

The type of resource data sync. If SyncType is SyncToDestination, then the resource data sync synchronizes data to an S3 bucket. If the SyncType is SyncFromSource then the resource data sync synchronizes data from Organizations or from multiple Amazon Web Services Regions.

s3Destination :: Maybe ResourceDataSyncS3Destination

Configuration information for the target S3 bucket.

syncLastModifiedTime :: Maybe POSIX

The date and time the resource data sync was changed.

syncName :: Maybe Text

The name of the resource data sync.

lastSyncTime :: Maybe POSIX

The last time the configuration attempted to sync (UTC).

syncSource :: Maybe ResourceDataSyncSourceWithState

Information about the source where the data was synchronized.

lastSuccessfulSyncTime :: Maybe POSIX

The last time the sync operations returned a status of SUCCESSFUL (UTC).

lastStatus :: Maybe LastResourceDataSyncStatus

The status reported by the last sync.

lastSyncStatusMessage :: Maybe Text

The status message details reported by the last sync.

syncCreatedTime :: Maybe POSIX

The date and time the configuration was created (UTC).

newResourceDataSyncItem :: ResourceDataSyncItem

Create a value of ResourceDataSyncItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:syncType:ResourceDataSyncItem', resourceDataSyncItem_syncType - The type of resource data sync. If SyncType is SyncToDestination, then the resource data sync synchronizes data to an S3 bucket. If the SyncType is SyncFromSource then the resource data sync synchronizes data from Organizations or from multiple Amazon Web Services Regions.

$sel:s3Destination:ResourceDataSyncItem', resourceDataSyncItem_s3Destination - Configuration information for the target S3 bucket.

$sel:syncLastModifiedTime:ResourceDataSyncItem', resourceDataSyncItem_syncLastModifiedTime - The date and time the resource data sync was changed.

$sel:syncName:ResourceDataSyncItem', resourceDataSyncItem_syncName - The name of the resource data sync.

$sel:lastSyncTime:ResourceDataSyncItem', resourceDataSyncItem_lastSyncTime - The last time the configuration attempted to sync (UTC).

$sel:syncSource:ResourceDataSyncItem', resourceDataSyncItem_syncSource - Information about the source where the data was synchronized.

$sel:lastSuccessfulSyncTime:ResourceDataSyncItem', resourceDataSyncItem_lastSuccessfulSyncTime - The last time the sync operations returned a status of SUCCESSFUL (UTC).

$sel:lastStatus:ResourceDataSyncItem', resourceDataSyncItem_lastStatus - The status reported by the last sync.

$sel:lastSyncStatusMessage:ResourceDataSyncItem', resourceDataSyncItem_lastSyncStatusMessage - The status message details reported by the last sync.

$sel:syncCreatedTime:ResourceDataSyncItem', resourceDataSyncItem_syncCreatedTime - The date and time the configuration was created (UTC).

resourceDataSyncItem_syncType :: Lens' ResourceDataSyncItem (Maybe Text)

The type of resource data sync. If SyncType is SyncToDestination, then the resource data sync synchronizes data to an S3 bucket. If the SyncType is SyncFromSource then the resource data sync synchronizes data from Organizations or from multiple Amazon Web Services Regions.

resourceDataSyncItem_s3Destination :: Lens' ResourceDataSyncItem (Maybe ResourceDataSyncS3Destination)

Configuration information for the target S3 bucket.

resourceDataSyncItem_syncLastModifiedTime :: Lens' ResourceDataSyncItem (Maybe UTCTime)

The date and time the resource data sync was changed.

resourceDataSyncItem_syncName :: Lens' ResourceDataSyncItem (Maybe Text)

The name of the resource data sync.

resourceDataSyncItem_lastSyncTime :: Lens' ResourceDataSyncItem (Maybe UTCTime)

The last time the configuration attempted to sync (UTC).

resourceDataSyncItem_syncSource :: Lens' ResourceDataSyncItem (Maybe ResourceDataSyncSourceWithState)

Information about the source where the data was synchronized.

resourceDataSyncItem_lastSuccessfulSyncTime :: Lens' ResourceDataSyncItem (Maybe UTCTime)

The last time the sync operations returned a status of SUCCESSFUL (UTC).

resourceDataSyncItem_lastStatus :: Lens' ResourceDataSyncItem (Maybe LastResourceDataSyncStatus)

The status reported by the last sync.

resourceDataSyncItem_lastSyncStatusMessage :: Lens' ResourceDataSyncItem (Maybe Text)

The status message details reported by the last sync.

resourceDataSyncItem_syncCreatedTime :: Lens' ResourceDataSyncItem (Maybe UTCTime)

The date and time the configuration was created (UTC).