Network.AWS.SSM.Types.ResourceDataSyncSourceWithState

Description

Documentation

data ResourceDataSyncSourceWithState

The data type name for including resource data sync state. There are four sync states:

OrganizationNotExists (Your organization doesn't exist)

NoPermissions (The system can't locate the service-linked role. This role is automatically created when a user creates a resource data sync in Amazon Web Services Systems Manager Explorer.)

InvalidOrganizationalUnit (You specified or selected an invalid unit in the resource data sync configuration.)

TrustedAccessDisabled (You disabled Systems Manager access in the organization in Organizations.)

See: newResourceDataSyncSourceWithState smart constructor.

Constructors

ResourceDataSyncSourceWithState'

Fields

state :: Maybe Text

The data type name for including resource data sync state. There are four sync states:

OrganizationNotExists: Your organization doesn't exist.

NoPermissions: The system can't locate the service-linked role. This role is automatically created when a user creates a resource data sync in Explorer.

InvalidOrganizationalUnit: You specified or selected an invalid unit in the resource data sync configuration.

TrustedAccessDisabled: You disabled Systems Manager access in the organization in Organizations.

enableAllOpsDataSources :: Maybe Bool

When you create a resource data sync, if you choose one of the Organizations options, then Systems Manager automatically enables all OpsData sources in the selected Amazon Web Services Regions for all Amazon Web Services accounts in your organization (or in the selected organization units). For more information, see About multiple account and Region resource data syncs in the Amazon Web Services Systems Manager User Guide.

includeFutureRegions :: Maybe Bool

Whether to automatically synchronize and aggregate data from new Amazon Web Services Regions when those Regions come online.

sourceType :: Maybe Text

The type of data source for the resource data sync. SourceType is either AwsOrganizations (if an organization is present in Organizations) or singleAccountMultiRegions.

awsOrganizationsSource :: Maybe ResourceDataSyncAwsOrganizationsSource

The field name in SyncSource for the ResourceDataSyncAwsOrganizationsSource type.

sourceRegions :: Maybe [Text]

The SyncSource Amazon Web Services Regions included in the resource data sync.

newResourceDataSyncSourceWithState :: ResourceDataSyncSourceWithState

Create a value of ResourceDataSyncSourceWithState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:ResourceDataSyncSourceWithState', resourceDataSyncSourceWithState_state - The data type name for including resource data sync state. There are four sync states:

OrganizationNotExists: Your organization doesn't exist.

NoPermissions: The system can't locate the service-linked role. This role is automatically created when a user creates a resource data sync in Explorer.

InvalidOrganizationalUnit: You specified or selected an invalid unit in the resource data sync configuration.

TrustedAccessDisabled: You disabled Systems Manager access in the organization in Organizations.

$sel:enableAllOpsDataSources:ResourceDataSyncSourceWithState', resourceDataSyncSourceWithState_enableAllOpsDataSources - When you create a resource data sync, if you choose one of the Organizations options, then Systems Manager automatically enables all OpsData sources in the selected Amazon Web Services Regions for all Amazon Web Services accounts in your organization (or in the selected organization units). For more information, see About multiple account and Region resource data syncs in the Amazon Web Services Systems Manager User Guide.

$sel:includeFutureRegions:ResourceDataSyncSourceWithState', resourceDataSyncSourceWithState_includeFutureRegions - Whether to automatically synchronize and aggregate data from new Amazon Web Services Regions when those Regions come online.

$sel:sourceType:ResourceDataSyncSourceWithState', resourceDataSyncSourceWithState_sourceType - The type of data source for the resource data sync. SourceType is either AwsOrganizations (if an organization is present in Organizations) or singleAccountMultiRegions.

$sel:awsOrganizationsSource:ResourceDataSyncSourceWithState', resourceDataSyncSourceWithState_awsOrganizationsSource - The field name in SyncSource for the ResourceDataSyncAwsOrganizationsSource type.

$sel:sourceRegions:ResourceDataSyncSourceWithState', resourceDataSyncSourceWithState_sourceRegions - The SyncSource Amazon Web Services Regions included in the resource data sync.

resourceDataSyncSourceWithState_state :: Lens' ResourceDataSyncSourceWithState (Maybe Text)

The data type name for including resource data sync state. There are four sync states:

OrganizationNotExists: Your organization doesn't exist.

NoPermissions: The system can't locate the service-linked role. This role is automatically created when a user creates a resource data sync in Explorer.

InvalidOrganizationalUnit: You specified or selected an invalid unit in the resource data sync configuration.

TrustedAccessDisabled: You disabled Systems Manager access in the organization in Organizations.

resourceDataSyncSourceWithState_enableAllOpsDataSources :: Lens' ResourceDataSyncSourceWithState (Maybe Bool)

When you create a resource data sync, if you choose one of the Organizations options, then Systems Manager automatically enables all OpsData sources in the selected Amazon Web Services Regions for all Amazon Web Services accounts in your organization (or in the selected organization units). For more information, see About multiple account and Region resource data syncs in the Amazon Web Services Systems Manager User Guide.

resourceDataSyncSourceWithState_includeFutureRegions :: Lens' ResourceDataSyncSourceWithState (Maybe Bool)

Whether to automatically synchronize and aggregate data from new Amazon Web Services Regions when those Regions come online.

resourceDataSyncSourceWithState_sourceType :: Lens' ResourceDataSyncSourceWithState (Maybe Text)

The type of data source for the resource data sync. SourceType is either AwsOrganizations (if an organization is present in Organizations) or singleAccountMultiRegions.

resourceDataSyncSourceWithState_awsOrganizationsSource :: Lens' ResourceDataSyncSourceWithState (Maybe ResourceDataSyncAwsOrganizationsSource)

The field name in SyncSource for the ResourceDataSyncAwsOrganizationsSource type.

resourceDataSyncSourceWithState_sourceRegions :: Lens' ResourceDataSyncSourceWithState (Maybe [Text])

The SyncSource Amazon Web Services Regions included in the resource data sync.