Network.AWS.SSM.Types.Runbook

Description

Documentation

data Runbook

Information about an Automation runbook used in a runbook workflow in Change Manager.

The Automation runbooks specified for the runbook workflow can't run until all required approvals for the change request have been received.

See: newRunbook smart constructor.

Constructors

Runbook'

Fields

maxErrors :: Maybe Text

The MaxErrors value specified by the user when the execution started, indicating the maximum number of errors that can occur during the operation before the updates are stopped or rolled back.

targets :: Maybe [Target]

A key-value mapping to target resources that the runbook operation performs tasks on. Required if you specify TargetParameterName.

targetLocations :: Maybe (NonEmpty TargetLocation)

Information about the Amazon Web Services Regions and Amazon Web Services accounts targeted by the current Runbook operation.

targetParameterName :: Maybe Text

The name of the parameter used as the target resource for the rate-controlled runbook workflow. Required if you specify Targets.

maxConcurrency :: Maybe Text

The MaxConcurrency value specified by the user when the operation started, indicating the maximum number of resources that the runbook operation can run on at the same time.

documentVersion :: Maybe Text

The version of the Automation runbook used in a runbook workflow.

parameters :: Maybe (HashMap Text [Text])

The key-value map of execution parameters, which were supplied when calling StartChangeRequestExecution.

documentName :: Text

The name of the Automation runbook used in a runbook workflow.

newRunbook

Arguments

:: Text
$sel:documentName:Runbook'

-> Runbook

Create a value of Runbook with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxErrors:Runbook', runbook_maxErrors - The MaxErrors value specified by the user when the execution started, indicating the maximum number of errors that can occur during the operation before the updates are stopped or rolled back.

$sel:targets:Runbook', runbook_targets - A key-value mapping to target resources that the runbook operation performs tasks on. Required if you specify TargetParameterName.

$sel:targetLocations:Runbook', runbook_targetLocations - Information about the Amazon Web Services Regions and Amazon Web Services accounts targeted by the current Runbook operation.

$sel:targetParameterName:Runbook', runbook_targetParameterName - The name of the parameter used as the target resource for the rate-controlled runbook workflow. Required if you specify Targets.

$sel:maxConcurrency:Runbook', runbook_maxConcurrency - The MaxConcurrency value specified by the user when the operation started, indicating the maximum number of resources that the runbook operation can run on at the same time.

$sel:documentVersion:Runbook', runbook_documentVersion - The version of the Automation runbook used in a runbook workflow.

$sel:parameters:Runbook', runbook_parameters - The key-value map of execution parameters, which were supplied when calling StartChangeRequestExecution.

$sel:documentName:Runbook', runbook_documentName - The name of the Automation runbook used in a runbook workflow.

runbook_maxErrors :: Lens' Runbook (Maybe Text)

The MaxErrors value specified by the user when the execution started, indicating the maximum number of errors that can occur during the operation before the updates are stopped or rolled back.

runbook_targets :: Lens' Runbook (Maybe [Target])

A key-value mapping to target resources that the runbook operation performs tasks on. Required if you specify TargetParameterName.

runbook_targetLocations :: Lens' Runbook (Maybe (NonEmpty TargetLocation))

Information about the Amazon Web Services Regions and Amazon Web Services accounts targeted by the current Runbook operation.

runbook_targetParameterName :: Lens' Runbook (Maybe Text)

The name of the parameter used as the target resource for the rate-controlled runbook workflow. Required if you specify Targets.

runbook_maxConcurrency :: Lens' Runbook (Maybe Text)

The MaxConcurrency value specified by the user when the operation started, indicating the maximum number of resources that the runbook operation can run on at the same time.

runbook_documentVersion :: Lens' Runbook (Maybe Text)

The version of the Automation runbook used in a runbook workflow.

runbook_parameters :: Lens' Runbook (Maybe (HashMap Text [Text]))

The key-value map of execution parameters, which were supplied when calling StartChangeRequestExecution.

runbook_documentName :: Lens' Runbook Text

The name of the Automation runbook used in a runbook workflow.