Network.AWS.SSM.Types.ServiceSetting

Description

Documentation

data ServiceSetting

The service setting data structure.

ServiceSetting is an account-level setting for an Amazon Web Services service. This setting defines how a user interacts with or uses a service or a feature of a service. For example, if an Amazon Web Services service charges money to the account based on feature or service usage, then the Amazon Web Services service team might create a default setting of "false". This means the user can't use this feature unless they change the setting to "true" and intentionally opt in for a paid feature.

Services map a SettingId object to a setting value. Amazon Web Services services teams define the default value for a SettingId. You can't create a new SettingId, but you can overwrite the default value if you have the ssm:UpdateServiceSetting permission for the setting. Use the UpdateServiceSetting API operation to change the default setting. Or, use the ResetServiceSetting to change the value back to the original value defined by the Amazon Web Services service team.

See: newServiceSetting smart constructor.

Constructors

ServiceSetting'

Fields

lastModifiedDate :: Maybe POSIX

The last time the service setting was modified.

status :: Maybe Text

The status of the service setting. The value can be Default, Customized or PendingUpdate.

Default: The current setting uses a default value provisioned by the Amazon Web Services service team.

Customized: The current setting use a custom value specified by the customer.

PendingUpdate: The current setting uses a default or custom value, but a setting change request is pending approval.

settingValue :: Maybe Text

The value of the service setting.

arn :: Maybe Text

The ARN of the service setting.

settingId :: Maybe Text

The ID of the service setting.

lastModifiedUser :: Maybe Text

The ARN of the last modified user. This field is populated only if the setting value was overwritten.

newServiceSetting :: ServiceSetting

Create a value of ServiceSetting with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:ServiceSetting', serviceSetting_lastModifiedDate - The last time the service setting was modified.

$sel:status:ServiceSetting', serviceSetting_status - The status of the service setting. The value can be Default, Customized or PendingUpdate.

Default: The current setting uses a default value provisioned by the Amazon Web Services service team.

Customized: The current setting use a custom value specified by the customer.

PendingUpdate: The current setting uses a default or custom value, but a setting change request is pending approval.

$sel:settingValue:ServiceSetting', serviceSetting_settingValue - The value of the service setting.

$sel:arn:ServiceSetting', serviceSetting_arn - The ARN of the service setting.

$sel:settingId:ServiceSetting', serviceSetting_settingId - The ID of the service setting.

$sel:lastModifiedUser:ServiceSetting', serviceSetting_lastModifiedUser - The ARN of the last modified user. This field is populated only if the setting value was overwritten.

serviceSetting_lastModifiedDate :: Lens' ServiceSetting (Maybe UTCTime)

The last time the service setting was modified.

serviceSetting_status :: Lens' ServiceSetting (Maybe Text)

The status of the service setting. The value can be Default, Customized or PendingUpdate.

Default: The current setting uses a default value provisioned by the Amazon Web Services service team.

Customized: The current setting use a custom value specified by the customer.

PendingUpdate: The current setting uses a default or custom value, but a setting change request is pending approval.

serviceSetting_settingValue :: Lens' ServiceSetting (Maybe Text)

The value of the service setting.

serviceSetting_arn :: Lens' ServiceSetting (Maybe Text)

The ARN of the service setting.

serviceSetting_settingId :: Lens' ServiceSetting (Maybe Text)

The ID of the service setting.

serviceSetting_lastModifiedUser :: Lens' ServiceSetting (Maybe Text)

The ARN of the last modified user. This field is populated only if the setting value was overwritten.