Network.AWS.SSM.Types.Session

Description

Documentation

data Session

Information about a Session Manager connection to an instance.

See: newSession smart constructor.

Constructors

Session'

Fields

status :: Maybe SessionStatus

The status of the session. For example, "Connected" or "Terminated".

outputUrl :: Maybe SessionManagerOutputUrl

Reserved for future use.

documentName :: Maybe Text

The name of the Session Manager SSM document used to define the parameters and plugin settings for the session. For example, SSM-SessionManagerRunShell.

endDate :: Maybe POSIX

The date and time, in ISO-8601 Extended format, when the session was terminated.

owner :: Maybe Text

The ID of the Amazon Web Services user account that started the session.

startDate :: Maybe POSIX

The date and time, in ISO-8601 Extended format, when the session began.

details :: Maybe Text

Reserved for future use.

sessionId :: Maybe Text

The ID of the session.

target :: Maybe Text

The instance that the Session Manager session connected to.

newSession :: Session

Create a value of Session with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Session', session_status - The status of the session. For example, "Connected" or "Terminated".

$sel:outputUrl:Session', session_outputUrl - Reserved for future use.

$sel:documentName:Session', session_documentName - The name of the Session Manager SSM document used to define the parameters and plugin settings for the session. For example, SSM-SessionManagerRunShell.

$sel:endDate:Session', session_endDate - The date and time, in ISO-8601 Extended format, when the session was terminated.

$sel:owner:Session', session_owner - The ID of the Amazon Web Services user account that started the session.

$sel:startDate:Session', session_startDate - The date and time, in ISO-8601 Extended format, when the session began.

$sel:details:Session', session_details - Reserved for future use.

$sel:sessionId:Session', session_sessionId - The ID of the session.

$sel:target:Session', session_target - The instance that the Session Manager session connected to.

session_status :: Lens' Session (Maybe SessionStatus)

The status of the session. For example, "Connected" or "Terminated".

session_outputUrl :: Lens' Session (Maybe SessionManagerOutputUrl)

Reserved for future use.

session_documentName :: Lens' Session (Maybe Text)

The name of the Session Manager SSM document used to define the parameters and plugin settings for the session. For example, SSM-SessionManagerRunShell.

session_endDate :: Lens' Session (Maybe UTCTime)

The date and time, in ISO-8601 Extended format, when the session was terminated.

session_owner :: Lens' Session (Maybe Text)

The ID of the Amazon Web Services user account that started the session.

session_startDate :: Lens' Session (Maybe UTCTime)

The date and time, in ISO-8601 Extended format, when the session began.

session_details :: Lens' Session (Maybe Text)

Reserved for future use.

session_sessionId :: Lens' Session (Maybe Text)

The ID of the session.

session_target :: Lens' Session (Maybe Text)

The instance that the Session Manager session connected to.