Network.AWS.SSM.Types.StepExecution

Description

Documentation

data StepExecution

Detailed information about an the execution state of an Automation step.

See: newStepExecution smart constructor.

Constructors

StepExecution'

Fields

outputs :: Maybe (HashMap Text [Text])

Returned values from the execution of the step.

onFailure :: Maybe Text

The action to take if the step fails. The default value is Abort.

failureMessage :: Maybe Text

If a step failed, this message explains why the execution failed.

response :: Maybe Text

A message associated with the response code for an execution.

executionEndTime :: Maybe POSIX

If a step has finished execution, this contains the time the execution ended. If the step hasn't yet concluded, this field isn't populated.

nextStep :: Maybe Text

The next step after the step succeeds.

isEnd :: Maybe Bool

The flag which can be used to end automation no matter whether the step succeeds or fails.

maxAttempts :: Maybe Int

The maximum number of tries to run the action of the step. The default value is 1.

failureDetails :: Maybe FailureDetails

Information about the Automation failure.

targets :: Maybe [Target]

The targets for the step execution.

executionStartTime :: Maybe POSIX

If a step has begun execution, this contains the time the step started. If the step is in Pending status, this field isn't populated.

targetLocation :: Maybe TargetLocation

The combination of Amazon Web Services Regions and Amazon Web Services accounts targeted by the current Automation execution.

overriddenParameters :: Maybe (HashMap Text [Text])

A user-specified list of parameters to override when running a step.

isCritical :: Maybe Bool

The flag which can be used to help decide whether the failure of current step leads to the Automation failure.

action :: Maybe Text

The action this step performs. The action determines the behavior of the step.

responseCode :: Maybe Text

The response code returned by the execution of the step.

stepStatus :: Maybe AutomationExecutionStatus

The execution status for this step.

timeoutSeconds :: Maybe Integer

The timeout seconds of the step.

validNextSteps :: Maybe [Text]

Strategies used when step fails, we support Continue and Abort. Abort will fail the automation when the step fails. Continue will ignore the failure of current step and allow automation to run the next step. With conditional branching, we add step:stepName to support the automation to go to another specific step.

stepExecutionId :: Maybe Text

The unique ID of a step execution.

stepName :: Maybe Text

The name of this execution step.

inputs :: Maybe (HashMap Text Text)

Fully-resolved values passed into the step before execution.

newStepExecution :: StepExecution

Create a value of StepExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outputs:StepExecution', stepExecution_outputs - Returned values from the execution of the step.

$sel:onFailure:StepExecution', stepExecution_onFailure - The action to take if the step fails. The default value is Abort.

$sel:failureMessage:StepExecution', stepExecution_failureMessage - If a step failed, this message explains why the execution failed.

$sel:response:StepExecution', stepExecution_response - A message associated with the response code for an execution.

$sel:executionEndTime:StepExecution', stepExecution_executionEndTime - If a step has finished execution, this contains the time the execution ended. If the step hasn't yet concluded, this field isn't populated.

$sel:nextStep:StepExecution', stepExecution_nextStep - The next step after the step succeeds.

$sel:isEnd:StepExecution', stepExecution_isEnd - The flag which can be used to end automation no matter whether the step succeeds or fails.

$sel:maxAttempts:StepExecution', stepExecution_maxAttempts - The maximum number of tries to run the action of the step. The default value is 1.

$sel:failureDetails:StepExecution', stepExecution_failureDetails - Information about the Automation failure.

$sel:targets:StepExecution', stepExecution_targets - The targets for the step execution.

$sel:executionStartTime:StepExecution', stepExecution_executionStartTime - If a step has begun execution, this contains the time the step started. If the step is in Pending status, this field isn't populated.

$sel:targetLocation:StepExecution', stepExecution_targetLocation - The combination of Amazon Web Services Regions and Amazon Web Services accounts targeted by the current Automation execution.

$sel:overriddenParameters:StepExecution', stepExecution_overriddenParameters - A user-specified list of parameters to override when running a step.

$sel:isCritical:StepExecution', stepExecution_isCritical - The flag which can be used to help decide whether the failure of current step leads to the Automation failure.

$sel:action:StepExecution', stepExecution_action - The action this step performs. The action determines the behavior of the step.

$sel:responseCode:StepExecution', stepExecution_responseCode - The response code returned by the execution of the step.

$sel:stepStatus:StepExecution', stepExecution_stepStatus - The execution status for this step.

$sel:timeoutSeconds:StepExecution', stepExecution_timeoutSeconds - The timeout seconds of the step.

$sel:validNextSteps:StepExecution', stepExecution_validNextSteps - Strategies used when step fails, we support Continue and Abort. Abort will fail the automation when the step fails. Continue will ignore the failure of current step and allow automation to run the next step. With conditional branching, we add step:stepName to support the automation to go to another specific step.

$sel:stepExecutionId:StepExecution', stepExecution_stepExecutionId - The unique ID of a step execution.

$sel:stepName:StepExecution', stepExecution_stepName - The name of this execution step.

$sel:inputs:StepExecution', stepExecution_inputs - Fully-resolved values passed into the step before execution.

stepExecution_outputs :: Lens' StepExecution (Maybe (HashMap Text [Text]))

Returned values from the execution of the step.

stepExecution_onFailure :: Lens' StepExecution (Maybe Text)

The action to take if the step fails. The default value is Abort.

stepExecution_failureMessage :: Lens' StepExecution (Maybe Text)

If a step failed, this message explains why the execution failed.

stepExecution_response :: Lens' StepExecution (Maybe Text)

A message associated with the response code for an execution.

stepExecution_executionEndTime :: Lens' StepExecution (Maybe UTCTime)

If a step has finished execution, this contains the time the execution ended. If the step hasn't yet concluded, this field isn't populated.

stepExecution_nextStep :: Lens' StepExecution (Maybe Text)

The next step after the step succeeds.

stepExecution_isEnd :: Lens' StepExecution (Maybe Bool)

The flag which can be used to end automation no matter whether the step succeeds or fails.

stepExecution_maxAttempts :: Lens' StepExecution (Maybe Int)

The maximum number of tries to run the action of the step. The default value is 1.

stepExecution_failureDetails :: Lens' StepExecution (Maybe FailureDetails)

Information about the Automation failure.

stepExecution_targets :: Lens' StepExecution (Maybe [Target])

The targets for the step execution.

stepExecution_executionStartTime :: Lens' StepExecution (Maybe UTCTime)

If a step has begun execution, this contains the time the step started. If the step is in Pending status, this field isn't populated.

stepExecution_targetLocation :: Lens' StepExecution (Maybe TargetLocation)

The combination of Amazon Web Services Regions and Amazon Web Services accounts targeted by the current Automation execution.

stepExecution_overriddenParameters :: Lens' StepExecution (Maybe (HashMap Text [Text]))

A user-specified list of parameters to override when running a step.

stepExecution_isCritical :: Lens' StepExecution (Maybe Bool)

The flag which can be used to help decide whether the failure of current step leads to the Automation failure.

stepExecution_action :: Lens' StepExecution (Maybe Text)

The action this step performs. The action determines the behavior of the step.

stepExecution_responseCode :: Lens' StepExecution (Maybe Text)

The response code returned by the execution of the step.

stepExecution_stepStatus :: Lens' StepExecution (Maybe AutomationExecutionStatus)

The execution status for this step.

stepExecution_timeoutSeconds :: Lens' StepExecution (Maybe Integer)

The timeout seconds of the step.

stepExecution_validNextSteps :: Lens' StepExecution (Maybe [Text])

Strategies used when step fails, we support Continue and Abort. Abort will fail the automation when the step fails. Continue will ignore the failure of current step and allow automation to run the next step. With conditional branching, we add step:stepName to support the automation to go to another specific step.

stepExecution_stepExecutionId :: Lens' StepExecution (Maybe Text)

The unique ID of a step execution.

stepExecution_stepName :: Lens' StepExecution (Maybe Text)

The name of this execution step.

stepExecution_inputs :: Lens' StepExecution (Maybe (HashMap Text Text))

Fully-resolved values passed into the step before execution.