Network.AWS.SSM.Types.Target

Description

Documentation

data Target

An array of search criteria that targets instances using a key-value pair that you specify.

One or more targets must be specified for maintenance window Run Command-type tasks. Depending on the task, targets are optional for other maintenance window task types (Automation, Lambda, and Step Functions). For more information about running tasks that don't specify targets, see Registering maintenance window tasks without targets in the Amazon Web Services Systems Manager User Guide.

Supported formats include the following.

Key=InstanceIds,Values=instance-id-1,instance-id-2,instance-id-3

Key=tag:my-tag-key,Values=my-tag-value-1,my-tag-value-2

Key=tag-key,Values=my-tag-key-1,my-tag-key-2

Run Command and Maintenance window targets only: Key=resource-groups:Name,Values=resource-group-name

Maintenance window targets only: Key=resource-groups:ResourceTypeFilters,Values=resource-type-1,resource-type-2

Automation targets only: Key=ResourceGroup;Values=resource-group-name

For example:

Key=InstanceIds,Values=i-02573cafcfEXAMPLE,i-0471e04240EXAMPLE,i-07782c72faEXAMPLE

Key=tag:CostCenter,Values=CostCenter1,CostCenter2,CostCenter3

Key=tag-key,Values=Name,Instance-Type,CostCenter

Run Command and Maintenance window targets only: Key=resource-groups:Name,Values=ProductionResourceGroup

This example demonstrates how to target all resources in the resource group ProductionResourceGroup in your maintenance window.

Maintenance window targets only: Key=resource-groups:ResourceTypeFilters,Values=AWS::EC2::INSTANCE,AWS::EC2::VPC

This example demonstrates how to target only Amazon Elastic Compute Cloud (Amazon EC2) instances and VPCs in your maintenance window.

Automation targets only: Key=ResourceGroup,Values=MyResourceGroup

State Manager association targets only: Key=InstanceIds,Values=*

This example demonstrates how to target all managed instances in the Amazon Web Services Region where the association was created.

For more information about how to send commands that target instances using Key,Value parameters, see Targeting multiple instances in the Amazon Web Services Systems Manager User Guide.

See: newTarget smart constructor.

Constructors

Target'

Fields

key :: Maybe Text

User-defined criteria for sending commands that target instances that meet the criteria.

values :: Maybe [Text]

User-defined criteria that maps to Key. For example, if you specified tag:ServerRole, you could specify value:WebServer to run a command on instances that include EC2 tags of ServerRole,WebServer.

Depending on the type of target, the maximum number of values for a key might be lower than the global maximum of 50.

newTarget :: Target

Create a value of Target with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Target', target_key - User-defined criteria for sending commands that target instances that meet the criteria.

$sel:values:Target', target_values - User-defined criteria that maps to Key. For example, if you specified tag:ServerRole, you could specify value:WebServer to run a command on instances that include EC2 tags of ServerRole,WebServer.

Depending on the type of target, the maximum number of values for a key might be lower than the global maximum of 50.

target_key :: Lens' Target (Maybe Text)

User-defined criteria for sending commands that target instances that meet the criteria.

target_values :: Lens' Target (Maybe [Text])

User-defined criteria that maps to Key. For example, if you specified tag:ServerRole, you could specify value:WebServer to run a command on instances that include EC2 tags of ServerRole,WebServer.

Depending on the type of target, the maximum number of values for a key might be lower than the global maximum of 50.