Network.AWS.SSM.Types.TargetLocation

Description

Documentation

data TargetLocation

The combination of Amazon Web Services Regions and Amazon Web Services accounts targeted by the current Automation execution.

See: newTargetLocation smart constructor.

Constructors

TargetLocation'

Fields

accounts :: Maybe (NonEmpty Text)

The Amazon Web Services accounts targeted by the current Automation execution.

targetLocationMaxConcurrency :: Maybe Text

The maximum number of Amazon Web Services Regions and Amazon Web Services accounts allowed to run the Automation concurrently.

targetLocationMaxErrors :: Maybe Text

The maximum number of errors allowed before the system stops queueing additional Automation executions for the currently running Automation.

regions :: Maybe (NonEmpty Text)

The Amazon Web Services Regions targeted by the current Automation execution.

executionRoleName :: Maybe Text

The Automation execution role used by the currently running Automation. If not specified, the default value is AWS-SystemsManager-AutomationExecutionRole.

newTargetLocation :: TargetLocation

Create a value of TargetLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accounts:TargetLocation', targetLocation_accounts - The Amazon Web Services accounts targeted by the current Automation execution.

$sel:targetLocationMaxConcurrency:TargetLocation', targetLocation_targetLocationMaxConcurrency - The maximum number of Amazon Web Services Regions and Amazon Web Services accounts allowed to run the Automation concurrently.

$sel:targetLocationMaxErrors:TargetLocation', targetLocation_targetLocationMaxErrors - The maximum number of errors allowed before the system stops queueing additional Automation executions for the currently running Automation.

$sel:regions:TargetLocation', targetLocation_regions - The Amazon Web Services Regions targeted by the current Automation execution.

$sel:executionRoleName:TargetLocation', targetLocation_executionRoleName - The Automation execution role used by the currently running Automation. If not specified, the default value is AWS-SystemsManager-AutomationExecutionRole.

targetLocation_accounts :: Lens' TargetLocation (Maybe (NonEmpty Text))

The Amazon Web Services accounts targeted by the current Automation execution.

targetLocation_targetLocationMaxConcurrency :: Lens' TargetLocation (Maybe Text)

The maximum number of Amazon Web Services Regions and Amazon Web Services accounts allowed to run the Automation concurrently.

targetLocation_targetLocationMaxErrors :: Lens' TargetLocation (Maybe Text)

The maximum number of errors allowed before the system stops queueing additional Automation executions for the currently running Automation.

targetLocation_regions :: Lens' TargetLocation (Maybe (NonEmpty Text))

The Amazon Web Services Regions targeted by the current Automation execution.

targetLocation_executionRoleName :: Lens' TargetLocation (Maybe Text)

The Automation execution role used by the currently running Automation. If not specified, the default value is AWS-SystemsManager-AutomationExecutionRole.