Network.AWS.SSM.UpdateAssociation

Description

Updates an association. You can update the association name and version, the document version, schedule, parameters, and Amazon Simple Storage Service (Amazon S3) output.

In order to call this API operation, your Identity and Access Management (IAM) user account, group, or role must be configured with permission to call the DescribeAssociation API operation. If you don't have permission to call DescribeAssociation, then you receive the following error: An error occurred (AccessDeniedException) when calling the UpdateAssociation operation: User: <user_arn> isn't authorized to perform: ssm:DescribeAssociation on resource: <resource_arn>

When you update an association, the association immediately runs against the specified targets.

Creating a Request

data UpdateAssociation

See: newUpdateAssociation smart constructor.

Constructors

UpdateAssociation'

Fields

targetLocations :: Maybe (NonEmpty TargetLocation)

A location is a combination of Amazon Web Services Regions and Amazon Web Services accounts where you want to run the association. Use this action to update an association in multiple Regions and multiple accounts.

applyOnlyAtCronInterval :: Maybe Bool

By default, when you update an association, the system runs it immediately after it is updated and then according to the schedule you specified. Specify this option if you don't want an association to run immediately after you update it. This parameter isn't supported for rate expressions.

Also, if you specified this option when you created the association, you can reset it. To do so, specify the no-apply-only-at-cron-interval parameter when you update the association from the command line. This parameter forces the association to run immediately after updating it and according to the interval specified.

maxErrors :: Maybe Text

The number of errors that are allowed before the system stops sending requests to run the association on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops sending requests when the fourth error is received. If you specify 0, then the system stops sending requests after the first error is returned. If you run an association on 50 instances and set MaxError to 10%, then the system stops sending the request when the sixth error is received.

Executions that are already running an association when MaxErrors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set MaxConcurrency to 1 so that executions proceed one at a time.

scheduleExpression :: Maybe Text

The cron expression used to schedule the association that you want to update.

name :: Maybe Text

The name of the SSM Command document or Automation runbook that contains the configuration information for the instance.

You can specify Amazon Web Services-predefined documents, documents you created, or a document that is shared with you from another account.

For Systems Manager document (SSM document) that are shared with you from other Amazon Web Services accounts, you must specify the complete SSM document ARN, in the following format:

arn:aws:ssm:region:account-id:document/document-name

For example:

arn:aws:ssm:us-east-2:12345678912:document/My-Shared-Document

For Amazon Web Services-predefined documents and SSM documents you created in your account, you only need to specify the document name. For example, AWS-ApplyPatchBaseline or My-Document.

outputLocation :: Maybe InstanceAssociationOutputLocation

An S3 bucket where you want to store the results of this request.

syncCompliance :: Maybe AssociationSyncCompliance

The mode for generating association compliance. You can specify AUTO or MANUAL. In AUTO mode, the system uses the status of the association execution to determine the compliance status. If the association execution runs successfully, then the association is COMPLIANT. If the association execution doesn't run successfully, the association is NON-COMPLIANT.

In MANUAL mode, you must specify the AssociationId as a parameter for the PutComplianceItems API operation. In this case, compliance data isn't managed by State Manager, a capability of Amazon Web Services Systems Manager. It is managed by your direct call to the PutComplianceItems API operation.

By default, all associations use AUTO mode.

targets :: Maybe [Target]

The targets of the association.

parameters :: Maybe (HashMap Text [Text])

The parameters you want to update for the association. If you create a parameter using Parameter Store, a capability of Amazon Web Services Systems Manager, you can reference the parameter using {{ssm:parameter-name}}.

documentVersion :: Maybe Text

The document version you want update for the association.

automationTargetParameterName :: Maybe Text

Specify the target for the association. This target is required for associations that use an Automation runbook and target resources by using rate controls. Automation is a capability of Amazon Web Services Systems Manager.

associationVersion :: Maybe Text

This parameter is provided for concurrency control purposes. You must specify the latest association version in the service. If you want to ensure that this request succeeds, either specify $LATEST, or omit this parameter.

associationName :: Maybe Text

The name of the association that you want to update.

calendarNames :: Maybe [Text]

The names or Amazon Resource Names (ARNs) of the Change Calendar type documents you want to gate your associations under. The associations only run when that change calendar is open. For more information, see Amazon Web Services Systems Manager Change Calendar.

complianceSeverity :: Maybe AssociationComplianceSeverity

The severity level to assign to the association.

maxConcurrency :: Maybe Text

The maximum number of targets allowed to run the association at the same time. You can specify a number, for example 10, or a percentage of the target set, for example 10%. The default value is 100%, which means all targets run the association at the same time.

If a new instance starts and attempts to run an association while Systems Manager is running MaxConcurrency associations, the association is allowed to run. During the next association interval, the new instance will process its association within the limit specified for MaxConcurrency.

associationId :: Text

The ID of the association you want to update.

newUpdateAssociation

Arguments

:: Text
$sel:associationId:UpdateAssociation'

-> UpdateAssociation

Create a value of UpdateAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetLocations:UpdateAssociation', updateAssociation_targetLocations - A location is a combination of Amazon Web Services Regions and Amazon Web Services accounts where you want to run the association. Use this action to update an association in multiple Regions and multiple accounts.

$sel:applyOnlyAtCronInterval:UpdateAssociation', updateAssociation_applyOnlyAtCronInterval - By default, when you update an association, the system runs it immediately after it is updated and then according to the schedule you specified. Specify this option if you don't want an association to run immediately after you update it. This parameter isn't supported for rate expressions.

Also, if you specified this option when you created the association, you can reset it. To do so, specify the no-apply-only-at-cron-interval parameter when you update the association from the command line. This parameter forces the association to run immediately after updating it and according to the interval specified.

$sel:maxErrors:UpdateAssociation', updateAssociation_maxErrors - The number of errors that are allowed before the system stops sending requests to run the association on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops sending requests when the fourth error is received. If you specify 0, then the system stops sending requests after the first error is returned. If you run an association on 50 instances and set MaxError to 10%, then the system stops sending the request when the sixth error is received.

Executions that are already running an association when MaxErrors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set MaxConcurrency to 1 so that executions proceed one at a time.

$sel:scheduleExpression:UpdateAssociation', updateAssociation_scheduleExpression - The cron expression used to schedule the association that you want to update.

$sel:name:UpdateAssociation', updateAssociation_name - The name of the SSM Command document or Automation runbook that contains the configuration information for the instance.

You can specify Amazon Web Services-predefined documents, documents you created, or a document that is shared with you from another account.

For Systems Manager document (SSM document) that are shared with you from other Amazon Web Services accounts, you must specify the complete SSM document ARN, in the following format:

arn:aws:ssm:region:account-id:document/document-name

For example:

arn:aws:ssm:us-east-2:12345678912:document/My-Shared-Document

For Amazon Web Services-predefined documents and SSM documents you created in your account, you only need to specify the document name. For example, AWS-ApplyPatchBaseline or My-Document.

$sel:outputLocation:UpdateAssociation', updateAssociation_outputLocation - An S3 bucket where you want to store the results of this request.

$sel:syncCompliance:UpdateAssociation', updateAssociation_syncCompliance - The mode for generating association compliance. You can specify AUTO or MANUAL. In AUTO mode, the system uses the status of the association execution to determine the compliance status. If the association execution runs successfully, then the association is COMPLIANT. If the association execution doesn't run successfully, the association is NON-COMPLIANT.

In MANUAL mode, you must specify the AssociationId as a parameter for the PutComplianceItems API operation. In this case, compliance data isn't managed by State Manager, a capability of Amazon Web Services Systems Manager. It is managed by your direct call to the PutComplianceItems API operation.

By default, all associations use AUTO mode.

$sel:targets:UpdateAssociation', updateAssociation_targets - The targets of the association.

$sel:parameters:UpdateAssociation', updateAssociation_parameters - The parameters you want to update for the association. If you create a parameter using Parameter Store, a capability of Amazon Web Services Systems Manager, you can reference the parameter using {{ssm:parameter-name}}.

$sel:documentVersion:UpdateAssociation', updateAssociation_documentVersion - The document version you want update for the association.

$sel:automationTargetParameterName:UpdateAssociation', updateAssociation_automationTargetParameterName - Specify the target for the association. This target is required for associations that use an Automation runbook and target resources by using rate controls. Automation is a capability of Amazon Web Services Systems Manager.

$sel:associationVersion:UpdateAssociation', updateAssociation_associationVersion - This parameter is provided for concurrency control purposes. You must specify the latest association version in the service. If you want to ensure that this request succeeds, either specify $LATEST, or omit this parameter.

$sel:associationName:UpdateAssociation', updateAssociation_associationName - The name of the association that you want to update.

$sel:calendarNames:UpdateAssociation', updateAssociation_calendarNames - The names or Amazon Resource Names (ARNs) of the Change Calendar type documents you want to gate your associations under. The associations only run when that change calendar is open. For more information, see Amazon Web Services Systems Manager Change Calendar.

$sel:complianceSeverity:UpdateAssociation', updateAssociation_complianceSeverity - The severity level to assign to the association.

$sel:maxConcurrency:UpdateAssociation', updateAssociation_maxConcurrency - The maximum number of targets allowed to run the association at the same time. You can specify a number, for example 10, or a percentage of the target set, for example 10%. The default value is 100%, which means all targets run the association at the same time.

If a new instance starts and attempts to run an association while Systems Manager is running MaxConcurrency associations, the association is allowed to run. During the next association interval, the new instance will process its association within the limit specified for MaxConcurrency.

$sel:associationId:UpdateAssociation', updateAssociation_associationId - The ID of the association you want to update.

Request Lenses

updateAssociation_targetLocations :: Lens' UpdateAssociation (Maybe (NonEmpty TargetLocation))

A location is a combination of Amazon Web Services Regions and Amazon Web Services accounts where you want to run the association. Use this action to update an association in multiple Regions and multiple accounts.

updateAssociation_applyOnlyAtCronInterval :: Lens' UpdateAssociation (Maybe Bool)

By default, when you update an association, the system runs it immediately after it is updated and then according to the schedule you specified. Specify this option if you don't want an association to run immediately after you update it. This parameter isn't supported for rate expressions.

Also, if you specified this option when you created the association, you can reset it. To do so, specify the no-apply-only-at-cron-interval parameter when you update the association from the command line. This parameter forces the association to run immediately after updating it and according to the interval specified.

updateAssociation_maxErrors :: Lens' UpdateAssociation (Maybe Text)

The number of errors that are allowed before the system stops sending requests to run the association on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops sending requests when the fourth error is received. If you specify 0, then the system stops sending requests after the first error is returned. If you run an association on 50 instances and set MaxError to 10%, then the system stops sending the request when the sixth error is received.

Executions that are already running an association when MaxErrors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set MaxConcurrency to 1 so that executions proceed one at a time.

updateAssociation_scheduleExpression :: Lens' UpdateAssociation (Maybe Text)

The cron expression used to schedule the association that you want to update.

updateAssociation_name :: Lens' UpdateAssociation (Maybe Text)

The name of the SSM Command document or Automation runbook that contains the configuration information for the instance.

You can specify Amazon Web Services-predefined documents, documents you created, or a document that is shared with you from another account.

For Systems Manager document (SSM document) that are shared with you from other Amazon Web Services accounts, you must specify the complete SSM document ARN, in the following format:

arn:aws:ssm:region:account-id:document/document-name

For example:

arn:aws:ssm:us-east-2:12345678912:document/My-Shared-Document

For Amazon Web Services-predefined documents and SSM documents you created in your account, you only need to specify the document name. For example, AWS-ApplyPatchBaseline or My-Document.

updateAssociation_outputLocation :: Lens' UpdateAssociation (Maybe InstanceAssociationOutputLocation)

An S3 bucket where you want to store the results of this request.

updateAssociation_syncCompliance :: Lens' UpdateAssociation (Maybe AssociationSyncCompliance)

The mode for generating association compliance. You can specify AUTO or MANUAL. In AUTO mode, the system uses the status of the association execution to determine the compliance status. If the association execution runs successfully, then the association is COMPLIANT. If the association execution doesn't run successfully, the association is NON-COMPLIANT.

In MANUAL mode, you must specify the AssociationId as a parameter for the PutComplianceItems API operation. In this case, compliance data isn't managed by State Manager, a capability of Amazon Web Services Systems Manager. It is managed by your direct call to the PutComplianceItems API operation.

By default, all associations use AUTO mode.

updateAssociation_targets :: Lens' UpdateAssociation (Maybe [Target])

The targets of the association.

updateAssociation_parameters :: Lens' UpdateAssociation (Maybe (HashMap Text [Text]))

The parameters you want to update for the association. If you create a parameter using Parameter Store, a capability of Amazon Web Services Systems Manager, you can reference the parameter using {{ssm:parameter-name}}.

updateAssociation_documentVersion :: Lens' UpdateAssociation (Maybe Text)

The document version you want update for the association.

updateAssociation_automationTargetParameterName :: Lens' UpdateAssociation (Maybe Text)

Specify the target for the association. This target is required for associations that use an Automation runbook and target resources by using rate controls. Automation is a capability of Amazon Web Services Systems Manager.

updateAssociation_associationVersion :: Lens' UpdateAssociation (Maybe Text)

This parameter is provided for concurrency control purposes. You must specify the latest association version in the service. If you want to ensure that this request succeeds, either specify $LATEST, or omit this parameter.

updateAssociation_associationName :: Lens' UpdateAssociation (Maybe Text)

The name of the association that you want to update.

updateAssociation_calendarNames :: Lens' UpdateAssociation (Maybe [Text])

The names or Amazon Resource Names (ARNs) of the Change Calendar type documents you want to gate your associations under. The associations only run when that change calendar is open. For more information, see Amazon Web Services Systems Manager Change Calendar.

updateAssociation_complianceSeverity :: Lens' UpdateAssociation (Maybe AssociationComplianceSeverity)

The severity level to assign to the association.

updateAssociation_maxConcurrency :: Lens' UpdateAssociation (Maybe Text)

The maximum number of targets allowed to run the association at the same time. You can specify a number, for example 10, or a percentage of the target set, for example 10%. The default value is 100%, which means all targets run the association at the same time.

If a new instance starts and attempts to run an association while Systems Manager is running MaxConcurrency associations, the association is allowed to run. During the next association interval, the new instance will process its association within the limit specified for MaxConcurrency.

updateAssociation_associationId :: Lens' UpdateAssociation Text

The ID of the association you want to update.

Destructuring the Response

data UpdateAssociationResponse

See: newUpdateAssociationResponse smart constructor.

Constructors

UpdateAssociationResponse'

Fields

associationDescription :: Maybe AssociationDescription

The description of the association that was updated.

httpStatus :: Int

The response's http status code.

newUpdateAssociationResponse

Arguments

:: Int
$sel:httpStatus:UpdateAssociationResponse'

-> UpdateAssociationResponse

Create a value of UpdateAssociationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associationDescription:UpdateAssociationResponse', updateAssociationResponse_associationDescription - The description of the association that was updated.

$sel:httpStatus:UpdateAssociationResponse', updateAssociationResponse_httpStatus - The response's http status code.

Response Lenses

updateAssociationResponse_associationDescription :: Lens' UpdateAssociationResponse (Maybe AssociationDescription)

The description of the association that was updated.

updateAssociationResponse_httpStatus :: Lens' UpdateAssociationResponse Int

The response's http status code.