Network.AWS.SSM.UpdateDocument

Description

Updates one or more values for an SSM document.

Creating a Request

data UpdateDocument

See: newUpdateDocument smart constructor.

Constructors

UpdateDocument'

Fields

targetType :: Maybe Text

Specify a new target type for the document.

versionName :: Maybe Text

An optional field specifying the version of the artifact you are updating with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document, and can't be changed.

documentFormat :: Maybe DocumentFormat

Specify the document format for the new document version. Systems Manager supports JSON and YAML documents. JSON is the default format.

documentVersion :: Maybe Text

The version of the document that you want to update. Currently, Systems Manager supports updating only the latest version of the document. You can specify the version number of the latest version or use the $LATEST variable.

displayName :: Maybe Text

The friendly name of the SSM document that you want to update. This value can differ for each version of the document. If you don't specify a value for this parameter in your request, the existing value is applied to the new document version.

attachments :: Maybe [AttachmentsSource]

A list of key-value pairs that describe attachments to a version of a document.

content :: Text

A valid JSON or YAML string.

name :: Text

The name of the SSM document that you want to update.

newUpdateDocument

Arguments

:: Text
$sel:content:UpdateDocument'

-> Text
$sel:name:UpdateDocument'

-> UpdateDocument

Create a value of UpdateDocument with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetType:UpdateDocument', updateDocument_targetType - Specify a new target type for the document.

$sel:versionName:UpdateDocument', updateDocument_versionName - An optional field specifying the version of the artifact you are updating with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document, and can't be changed.

$sel:documentFormat:UpdateDocument', updateDocument_documentFormat - Specify the document format for the new document version. Systems Manager supports JSON and YAML documents. JSON is the default format.

$sel:documentVersion:UpdateDocument', updateDocument_documentVersion - The version of the document that you want to update. Currently, Systems Manager supports updating only the latest version of the document. You can specify the version number of the latest version or use the $LATEST variable.

$sel:displayName:UpdateDocument', updateDocument_displayName - The friendly name of the SSM document that you want to update. This value can differ for each version of the document. If you don't specify a value for this parameter in your request, the existing value is applied to the new document version.

$sel:attachments:UpdateDocument', updateDocument_attachments - A list of key-value pairs that describe attachments to a version of a document.

$sel:content:UpdateDocument', updateDocument_content - A valid JSON or YAML string.

$sel:name:UpdateDocument', updateDocument_name - The name of the SSM document that you want to update.

Request Lenses

updateDocument_targetType :: Lens' UpdateDocument (Maybe Text)

Specify a new target type for the document.

updateDocument_versionName :: Lens' UpdateDocument (Maybe Text)

An optional field specifying the version of the artifact you are updating with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document, and can't be changed.

updateDocument_documentFormat :: Lens' UpdateDocument (Maybe DocumentFormat)

Specify the document format for the new document version. Systems Manager supports JSON and YAML documents. JSON is the default format.

updateDocument_documentVersion :: Lens' UpdateDocument (Maybe Text)

The version of the document that you want to update. Currently, Systems Manager supports updating only the latest version of the document. You can specify the version number of the latest version or use the $LATEST variable.

updateDocument_displayName :: Lens' UpdateDocument (Maybe Text)

The friendly name of the SSM document that you want to update. This value can differ for each version of the document. If you don't specify a value for this parameter in your request, the existing value is applied to the new document version.

updateDocument_attachments :: Lens' UpdateDocument (Maybe [AttachmentsSource])

A list of key-value pairs that describe attachments to a version of a document.

updateDocument_content :: Lens' UpdateDocument Text

A valid JSON or YAML string.

updateDocument_name :: Lens' UpdateDocument Text

The name of the SSM document that you want to update.

Destructuring the Response

data UpdateDocumentResponse

See: newUpdateDocumentResponse smart constructor.

Constructors

UpdateDocumentResponse'

Fields

documentDescription :: Maybe DocumentDescription

A description of the document that was updated.

httpStatus :: Int

The response's http status code.

newUpdateDocumentResponse

Arguments

:: Int
$sel:httpStatus:UpdateDocumentResponse'

-> UpdateDocumentResponse

Create a value of UpdateDocumentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:documentDescription:UpdateDocumentResponse', updateDocumentResponse_documentDescription - A description of the document that was updated.

$sel:httpStatus:UpdateDocumentResponse', updateDocumentResponse_httpStatus - The response's http status code.

Response Lenses

updateDocumentResponse_documentDescription :: Lens' UpdateDocumentResponse (Maybe DocumentDescription)

A description of the document that was updated.

updateDocumentResponse_httpStatus :: Lens' UpdateDocumentResponse Int

The response's http status code.