Network.AWS.SSM.UpdateMaintenanceWindow

Description

Updates an existing maintenance window. Only specified parameters are modified.

The value you specify for Duration determines the specific end time for the maintenance window based on the time it begins. No maintenance window tasks are permitted to start after the resulting endtime minus the number of hours you specify for Cutoff. For example, if the maintenance window starts at 3 PM, the duration is three hours, and the value you specify for Cutoff is one hour, no maintenance window tasks can start after 5 PM.

Creating a Request

data UpdateMaintenanceWindow

See: newUpdateMaintenanceWindow smart constructor.

Constructors

UpdateMaintenanceWindow'

Fields

startDate :: Maybe Text

The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format. For example: "America/Los_Angeles", "UTC", or "Asia/Seoul". For more information, see the Time Zone Database on the IANA website.

duration :: Maybe Natural

The duration of the maintenance window in hours.

scheduleOffset :: Maybe Natural

The number of days to wait after the date and time specified by a cron expression before running the maintenance window.

For example, the following cron expression schedules a maintenance window to run the third Tuesday of every month at 11:30 PM.

cron(30 23 ? * TUE#3 *)

If the schedule offset is 2, the maintenance window won't run until two days later.

enabled :: Maybe Bool

Whether the maintenance window is enabled.

cutoff :: Maybe Natural

The number of hours before the end of the maintenance window that Amazon Web Services Systems Manager stops scheduling new tasks for execution.

name :: Maybe Text

The name of the maintenance window.

replace :: Maybe Bool

If True, then all fields that are required by the CreateMaintenanceWindow operation are also required for this API request. Optional fields that aren't specified are set to null.

scheduleTimezone :: Maybe Text

The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format. For example: "America/Los_Angeles", "UTC", or "Asia/Seoul". For more information, see the Time Zone Database on the IANA website.

description :: Maybe (Sensitive Text)

An optional description for the update request.

endDate :: Maybe Text

The date and time, in ISO-8601 Extended format, for when you want the maintenance window to become inactive. EndDate allows you to set a date and time in the future when the maintenance window will no longer run.

allowUnassociatedTargets :: Maybe Bool

Whether targets must be registered with the maintenance window before tasks can be defined for those targets.

schedule :: Maybe Text

The schedule of the maintenance window in the form of a cron or rate expression.

windowId :: Text

The ID of the maintenance window to update.

newUpdateMaintenanceWindow

Arguments

:: Text
$sel:windowId:UpdateMaintenanceWindow'

-> UpdateMaintenanceWindow

Create a value of UpdateMaintenanceWindow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startDate:UpdateMaintenanceWindow', updateMaintenanceWindow_startDate - The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format. For example: "America/Los_Angeles", "UTC", or "Asia/Seoul". For more information, see the Time Zone Database on the IANA website.

$sel:duration:UpdateMaintenanceWindow', updateMaintenanceWindow_duration - The duration of the maintenance window in hours.

$sel:scheduleOffset:UpdateMaintenanceWindow', updateMaintenanceWindow_scheduleOffset - The number of days to wait after the date and time specified by a cron expression before running the maintenance window.

For example, the following cron expression schedules a maintenance window to run the third Tuesday of every month at 11:30 PM.

cron(30 23 ? * TUE#3 *)

If the schedule offset is 2, the maintenance window won't run until two days later.

$sel:enabled:UpdateMaintenanceWindow', updateMaintenanceWindow_enabled - Whether the maintenance window is enabled.

$sel:cutoff:UpdateMaintenanceWindow', updateMaintenanceWindow_cutoff - The number of hours before the end of the maintenance window that Amazon Web Services Systems Manager stops scheduling new tasks for execution.

$sel:name:UpdateMaintenanceWindow', updateMaintenanceWindow_name - The name of the maintenance window.

$sel:replace:UpdateMaintenanceWindow', updateMaintenanceWindow_replace - If True, then all fields that are required by the CreateMaintenanceWindow operation are also required for this API request. Optional fields that aren't specified are set to null.

$sel:scheduleTimezone:UpdateMaintenanceWindow', updateMaintenanceWindow_scheduleTimezone - The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format. For example: "America/Los_Angeles", "UTC", or "Asia/Seoul". For more information, see the Time Zone Database on the IANA website.

$sel:description:UpdateMaintenanceWindow', updateMaintenanceWindow_description - An optional description for the update request.

$sel:endDate:UpdateMaintenanceWindow', updateMaintenanceWindow_endDate - The date and time, in ISO-8601 Extended format, for when you want the maintenance window to become inactive. EndDate allows you to set a date and time in the future when the maintenance window will no longer run.

$sel:allowUnassociatedTargets:UpdateMaintenanceWindow', updateMaintenanceWindow_allowUnassociatedTargets - Whether targets must be registered with the maintenance window before tasks can be defined for those targets.

$sel:schedule:UpdateMaintenanceWindow', updateMaintenanceWindow_schedule - The schedule of the maintenance window in the form of a cron or rate expression.

$sel:windowId:UpdateMaintenanceWindow', updateMaintenanceWindow_windowId - The ID of the maintenance window to update.

Request Lenses

updateMaintenanceWindow_startDate :: Lens' UpdateMaintenanceWindow (Maybe Text)

The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format. For example: "America/Los_Angeles", "UTC", or "Asia/Seoul". For more information, see the Time Zone Database on the IANA website.

updateMaintenanceWindow_duration :: Lens' UpdateMaintenanceWindow (Maybe Natural)

The duration of the maintenance window in hours.

updateMaintenanceWindow_scheduleOffset :: Lens' UpdateMaintenanceWindow (Maybe Natural)

The number of days to wait after the date and time specified by a cron expression before running the maintenance window.

For example, the following cron expression schedules a maintenance window to run the third Tuesday of every month at 11:30 PM.

cron(30 23 ? * TUE#3 *)

If the schedule offset is 2, the maintenance window won't run until two days later.

updateMaintenanceWindow_enabled :: Lens' UpdateMaintenanceWindow (Maybe Bool)

Whether the maintenance window is enabled.

updateMaintenanceWindow_cutoff :: Lens' UpdateMaintenanceWindow (Maybe Natural)

The number of hours before the end of the maintenance window that Amazon Web Services Systems Manager stops scheduling new tasks for execution.

updateMaintenanceWindow_name :: Lens' UpdateMaintenanceWindow (Maybe Text)

The name of the maintenance window.

updateMaintenanceWindow_replace :: Lens' UpdateMaintenanceWindow (Maybe Bool)

If True, then all fields that are required by the CreateMaintenanceWindow operation are also required for this API request. Optional fields that aren't specified are set to null.

updateMaintenanceWindow_scheduleTimezone :: Lens' UpdateMaintenanceWindow (Maybe Text)

The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format. For example: "America/Los_Angeles", "UTC", or "Asia/Seoul". For more information, see the Time Zone Database on the IANA website.

updateMaintenanceWindow_description :: Lens' UpdateMaintenanceWindow (Maybe Text)

An optional description for the update request.

updateMaintenanceWindow_endDate :: Lens' UpdateMaintenanceWindow (Maybe Text)

The date and time, in ISO-8601 Extended format, for when you want the maintenance window to become inactive. EndDate allows you to set a date and time in the future when the maintenance window will no longer run.

updateMaintenanceWindow_allowUnassociatedTargets :: Lens' UpdateMaintenanceWindow (Maybe Bool)

Whether targets must be registered with the maintenance window before tasks can be defined for those targets.

updateMaintenanceWindow_schedule :: Lens' UpdateMaintenanceWindow (Maybe Text)

The schedule of the maintenance window in the form of a cron or rate expression.

updateMaintenanceWindow_windowId :: Lens' UpdateMaintenanceWindow Text

The ID of the maintenance window to update.

Destructuring the Response

data UpdateMaintenanceWindowResponse

See: newUpdateMaintenanceWindowResponse smart constructor.

Constructors

UpdateMaintenanceWindowResponse'

Fields

startDate :: Maybe Text

The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become active. The maintenance window won't run before this specified time.

duration :: Maybe Natural

The duration of the maintenance window in hours.

scheduleOffset :: Maybe Natural

The number of days to wait to run a maintenance window after the scheduled cron expression date and time.

enabled :: Maybe Bool

Whether the maintenance window is enabled.

cutoff :: Maybe Natural

The number of hours before the end of the maintenance window that Amazon Web Services Systems Manager stops scheduling new tasks for execution.

name :: Maybe Text

The name of the maintenance window.

windowId :: Maybe Text

The ID of the created maintenance window.

scheduleTimezone :: Maybe Text

The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format. For example: "America/Los_Angeles", "UTC", or "Asia/Seoul". For more information, see the Time Zone Database on the IANA website.

description :: Maybe (Sensitive Text)

An optional description of the update.

endDate :: Maybe Text

The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become inactive. The maintenance window won't run after this specified time.

allowUnassociatedTargets :: Maybe Bool

Whether targets must be registered with the maintenance window before tasks can be defined for those targets.

schedule :: Maybe Text

The schedule of the maintenance window in the form of a cron or rate expression.

httpStatus :: Int

The response's http status code.

newUpdateMaintenanceWindowResponse

Arguments

:: Int
$sel:httpStatus:UpdateMaintenanceWindowResponse'

-> UpdateMaintenanceWindowResponse

Create a value of UpdateMaintenanceWindowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startDate:UpdateMaintenanceWindow', updateMaintenanceWindowResponse_startDate - The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become active. The maintenance window won't run before this specified time.

$sel:duration:UpdateMaintenanceWindow', updateMaintenanceWindowResponse_duration - The duration of the maintenance window in hours.

$sel:scheduleOffset:UpdateMaintenanceWindow', updateMaintenanceWindowResponse_scheduleOffset - The number of days to wait to run a maintenance window after the scheduled cron expression date and time.

$sel:enabled:UpdateMaintenanceWindow', updateMaintenanceWindowResponse_enabled - Whether the maintenance window is enabled.

$sel:cutoff:UpdateMaintenanceWindow', updateMaintenanceWindowResponse_cutoff - The number of hours before the end of the maintenance window that Amazon Web Services Systems Manager stops scheduling new tasks for execution.

$sel:name:UpdateMaintenanceWindow', updateMaintenanceWindowResponse_name - The name of the maintenance window.

$sel:windowId:UpdateMaintenanceWindow', updateMaintenanceWindowResponse_windowId - The ID of the created maintenance window.

$sel:scheduleTimezone:UpdateMaintenanceWindow', updateMaintenanceWindowResponse_scheduleTimezone - The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format. For example: "America/Los_Angeles", "UTC", or "Asia/Seoul". For more information, see the Time Zone Database on the IANA website.

$sel:description:UpdateMaintenanceWindow', updateMaintenanceWindowResponse_description - An optional description of the update.

$sel:endDate:UpdateMaintenanceWindow', updateMaintenanceWindowResponse_endDate - The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become inactive. The maintenance window won't run after this specified time.

$sel:allowUnassociatedTargets:UpdateMaintenanceWindow', updateMaintenanceWindowResponse_allowUnassociatedTargets - Whether targets must be registered with the maintenance window before tasks can be defined for those targets.

$sel:schedule:UpdateMaintenanceWindow', updateMaintenanceWindowResponse_schedule - The schedule of the maintenance window in the form of a cron or rate expression.

$sel:httpStatus:UpdateMaintenanceWindowResponse', updateMaintenanceWindowResponse_httpStatus - The response's http status code.

Response Lenses

updateMaintenanceWindowResponse_startDate :: Lens' UpdateMaintenanceWindowResponse (Maybe Text)

The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become active. The maintenance window won't run before this specified time.

updateMaintenanceWindowResponse_duration :: Lens' UpdateMaintenanceWindowResponse (Maybe Natural)

The duration of the maintenance window in hours.

updateMaintenanceWindowResponse_scheduleOffset :: Lens' UpdateMaintenanceWindowResponse (Maybe Natural)

The number of days to wait to run a maintenance window after the scheduled cron expression date and time.

updateMaintenanceWindowResponse_enabled :: Lens' UpdateMaintenanceWindowResponse (Maybe Bool)

Whether the maintenance window is enabled.

updateMaintenanceWindowResponse_cutoff :: Lens' UpdateMaintenanceWindowResponse (Maybe Natural)

The number of hours before the end of the maintenance window that Amazon Web Services Systems Manager stops scheduling new tasks for execution.

updateMaintenanceWindowResponse_name :: Lens' UpdateMaintenanceWindowResponse (Maybe Text)

The name of the maintenance window.

updateMaintenanceWindowResponse_windowId :: Lens' UpdateMaintenanceWindowResponse (Maybe Text)

The ID of the created maintenance window.

updateMaintenanceWindowResponse_scheduleTimezone :: Lens' UpdateMaintenanceWindowResponse (Maybe Text)

The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format. For example: "America/Los_Angeles", "UTC", or "Asia/Seoul". For more information, see the Time Zone Database on the IANA website.

updateMaintenanceWindowResponse_description :: Lens' UpdateMaintenanceWindowResponse (Maybe Text)

An optional description of the update.

updateMaintenanceWindowResponse_endDate :: Lens' UpdateMaintenanceWindowResponse (Maybe Text)

The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become inactive. The maintenance window won't run after this specified time.

updateMaintenanceWindowResponse_allowUnassociatedTargets :: Lens' UpdateMaintenanceWindowResponse (Maybe Bool)

Whether targets must be registered with the maintenance window before tasks can be defined for those targets.

updateMaintenanceWindowResponse_schedule :: Lens' UpdateMaintenanceWindowResponse (Maybe Text)

The schedule of the maintenance window in the form of a cron or rate expression.

updateMaintenanceWindowResponse_httpStatus :: Lens' UpdateMaintenanceWindowResponse Int

The response's http status code.