Network.AWS.SSM.UpdateMaintenanceWindowTarget

Description

Modifies the target of an existing maintenance window. You can change the following:

Name

Description

Owner

IDs for an ID target

Tags for a Tag target

From any supported tag type to another. The three supported tag types are ID target, Tag target, and resource group. For more information, see Target.

If a parameter is null, then the corresponding field isn't modified.

Creating a Request

data UpdateMaintenanceWindowTarget

See: newUpdateMaintenanceWindowTarget smart constructor.

Constructors

UpdateMaintenanceWindowTarget'

Fields

targets :: Maybe [Target]

The targets to add or replace.

name :: Maybe Text

A name for the update.

replace :: Maybe Bool

If True, then all fields that are required by the RegisterTargetWithMaintenanceWindow operation are also required for this API request. Optional fields that aren't specified are set to null.

description :: Maybe (Sensitive Text)

An optional description for the update.

ownerInformation :: Maybe (Sensitive Text)

User-provided value that will be included in any Amazon CloudWatch Events events raised while running tasks for these targets in this maintenance window.

windowId :: Text

The maintenance window ID with which to modify the target.

windowTargetId :: Text

The target ID to modify.

newUpdateMaintenanceWindowTarget

Arguments

:: Text
$sel:windowId:UpdateMaintenanceWindowTarget'

-> Text
$sel:windowTargetId:UpdateMaintenanceWindowTarget'

-> UpdateMaintenanceWindowTarget

Create a value of UpdateMaintenanceWindowTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targets:UpdateMaintenanceWindowTarget', updateMaintenanceWindowTarget_targets - The targets to add or replace.

$sel:name:UpdateMaintenanceWindowTarget', updateMaintenanceWindowTarget_name - A name for the update.

$sel:replace:UpdateMaintenanceWindowTarget', updateMaintenanceWindowTarget_replace - If True, then all fields that are required by the RegisterTargetWithMaintenanceWindow operation are also required for this API request. Optional fields that aren't specified are set to null.

$sel:description:UpdateMaintenanceWindowTarget', updateMaintenanceWindowTarget_description - An optional description for the update.

$sel:ownerInformation:UpdateMaintenanceWindowTarget', updateMaintenanceWindowTarget_ownerInformation - User-provided value that will be included in any Amazon CloudWatch Events events raised while running tasks for these targets in this maintenance window.

$sel:windowId:UpdateMaintenanceWindowTarget', updateMaintenanceWindowTarget_windowId - The maintenance window ID with which to modify the target.

$sel:windowTargetId:UpdateMaintenanceWindowTarget', updateMaintenanceWindowTarget_windowTargetId - The target ID to modify.

Request Lenses

updateMaintenanceWindowTarget_targets :: Lens' UpdateMaintenanceWindowTarget (Maybe [Target])

The targets to add or replace.

updateMaintenanceWindowTarget_name :: Lens' UpdateMaintenanceWindowTarget (Maybe Text)

A name for the update.

updateMaintenanceWindowTarget_replace :: Lens' UpdateMaintenanceWindowTarget (Maybe Bool)

If True, then all fields that are required by the RegisterTargetWithMaintenanceWindow operation are also required for this API request. Optional fields that aren't specified are set to null.

updateMaintenanceWindowTarget_description :: Lens' UpdateMaintenanceWindowTarget (Maybe Text)

An optional description for the update.

updateMaintenanceWindowTarget_ownerInformation :: Lens' UpdateMaintenanceWindowTarget (Maybe Text)

User-provided value that will be included in any Amazon CloudWatch Events events raised while running tasks for these targets in this maintenance window.

updateMaintenanceWindowTarget_windowId :: Lens' UpdateMaintenanceWindowTarget Text

The maintenance window ID with which to modify the target.

updateMaintenanceWindowTarget_windowTargetId :: Lens' UpdateMaintenanceWindowTarget Text

The target ID to modify.

Destructuring the Response

data UpdateMaintenanceWindowTargetResponse

See: newUpdateMaintenanceWindowTargetResponse smart constructor.

Constructors

UpdateMaintenanceWindowTargetResponse'

Fields

windowTargetId :: Maybe Text

The target ID specified in the update request.

targets :: Maybe [Target]

The updated targets.

name :: Maybe Text

The updated name.

windowId :: Maybe Text

The maintenance window ID specified in the update request.

description :: Maybe (Sensitive Text)

The updated description.

ownerInformation :: Maybe (Sensitive Text)

The updated owner.

httpStatus :: Int

The response's http status code.

newUpdateMaintenanceWindowTargetResponse

Arguments

:: Int
$sel:httpStatus:UpdateMaintenanceWindowTargetResponse'

-> UpdateMaintenanceWindowTargetResponse

Create a value of UpdateMaintenanceWindowTargetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:windowTargetId:UpdateMaintenanceWindowTarget', updateMaintenanceWindowTargetResponse_windowTargetId - The target ID specified in the update request.

$sel:targets:UpdateMaintenanceWindowTarget', updateMaintenanceWindowTargetResponse_targets - The updated targets.

$sel:name:UpdateMaintenanceWindowTarget', updateMaintenanceWindowTargetResponse_name - The updated name.

$sel:windowId:UpdateMaintenanceWindowTarget', updateMaintenanceWindowTargetResponse_windowId - The maintenance window ID specified in the update request.

$sel:description:UpdateMaintenanceWindowTarget', updateMaintenanceWindowTargetResponse_description - The updated description.

$sel:ownerInformation:UpdateMaintenanceWindowTarget', updateMaintenanceWindowTargetResponse_ownerInformation - The updated owner.

$sel:httpStatus:UpdateMaintenanceWindowTargetResponse', updateMaintenanceWindowTargetResponse_httpStatus - The response's http status code.

Response Lenses

updateMaintenanceWindowTargetResponse_windowTargetId :: Lens' UpdateMaintenanceWindowTargetResponse (Maybe Text)

The target ID specified in the update request.

updateMaintenanceWindowTargetResponse_targets :: Lens' UpdateMaintenanceWindowTargetResponse (Maybe [Target])

The updated targets.

updateMaintenanceWindowTargetResponse_name :: Lens' UpdateMaintenanceWindowTargetResponse (Maybe Text)

The updated name.

updateMaintenanceWindowTargetResponse_windowId :: Lens' UpdateMaintenanceWindowTargetResponse (Maybe Text)

The maintenance window ID specified in the update request.

updateMaintenanceWindowTargetResponse_description :: Lens' UpdateMaintenanceWindowTargetResponse (Maybe Text)

The updated description.

updateMaintenanceWindowTargetResponse_ownerInformation :: Lens' UpdateMaintenanceWindowTargetResponse (Maybe Text)

The updated owner.

updateMaintenanceWindowTargetResponse_httpStatus :: Lens' UpdateMaintenanceWindowTargetResponse Int

The response's http status code.