Network.AWS.SSM.UpdateMaintenanceWindowTask

Description

Modifies a task assigned to a maintenance window. You can't change the task type, but you can change the following values:

TaskARN. For example, you can change a RUN_COMMAND task from AWS-RunPowerShellScript to AWS-RunShellScript.

ServiceRoleArn

TaskInvocationParameters

Priority

MaxConcurrency

MaxErrors

One or more targets must be specified for maintenance window Run Command-type tasks. Depending on the task, targets are optional for other maintenance window task types (Automation, Lambda, and Step Functions). For more information about running tasks that don't specify targets, see Registering maintenance window tasks without targets in the Amazon Web Services Systems Manager User Guide.

If the value for a parameter in UpdateMaintenanceWindowTask is null, then the corresponding field isn't modified. If you set Replace to true, then all fields required by the RegisterTaskWithMaintenanceWindow operation are required for this request. Optional fields that aren't specified are set to null.

When you update a maintenance window task that has options specified in TaskInvocationParameters, you must provide again all the TaskInvocationParameters values that you want to retain. The values you don't specify again are removed. For example, suppose that when you registered a Run Command task, you specified TaskInvocationParameters values for Comment, NotificationConfig, and OutputS3BucketName. If you update the maintenance window task and specify only a different OutputS3BucketName value, the values for Comment and NotificationConfig are removed.

Creating a Request

data UpdateMaintenanceWindowTask

See: newUpdateMaintenanceWindowTask smart constructor.

Constructors

UpdateMaintenanceWindowTask'

Fields

serviceRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM service role for Amazon Web Services Systems Manager to assume when running a maintenance window task. If you do not specify a service role ARN, Systems Manager uses your account's service-linked role. If no service-linked role for Systems Manager exists in your account, it is created when you run RegisterTaskWithMaintenanceWindow.

For more information, see the following topics in the in the /Amazon Web Services Systems Manager User Guide/:

Using service-linked roles for Systems Manager

Should I use a service-linked role or a custom service role to run maintenance window tasks?

replace :: Maybe Bool

If True, then all fields that are required by the RegisterTaskWithMaintenanceWindow operation are also required for this API request. Optional fields that aren't specified are set to null.

taskParameters :: Maybe (Sensitive (HashMap Text (Sensitive MaintenanceWindowTaskParameterValueExpression)))

The parameters to modify.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

The map has the following format:

Key: string, between 1 and 255 characters

Value: an array of strings, each string is between 1 and 255 characters

priority :: Maybe Natural

The new task priority to specify. The lower the number, the higher the priority. Tasks that have the same priority are scheduled in parallel.

taskArn :: Maybe Text

The task ARN to modify.

cutoffBehavior :: Maybe MaintenanceWindowTaskCutoffBehavior

Indicates whether tasks should continue to run after the cutoff time specified in the maintenance windows is reached.

CONTINUE_TASK: When the cutoff time is reached, any tasks that are running continue. The default value.

CANCEL_TASK:

For Automation, Lambda, Step Functions tasks: When the cutoff time is reached, any task invocations that are already running continue, but no new task invocations are started.

For Run Command tasks: When the cutoff time is reached, the system sends a CancelCommand operation that attempts to cancel the command associated with the task. However, there is no guarantee that the command will be terminated and the underlying process stopped.

The status for tasks that are not completed is TIMED_OUT.

maxErrors :: Maybe Text

The new MaxErrors value to specify. MaxErrors is the maximum number of errors that are allowed before the task stops being scheduled.

For maintenance window tasks without a target specified, you can't supply a value for this option. Instead, the system inserts a placeholder value of 1, which may be reported in the response to this command. This value doesn't affect the running of your task and can be ignored.

taskInvocationParameters :: Maybe MaintenanceWindowTaskInvocationParameters

The parameters that the task should use during execution. Populate only the fields that match the task type. All other fields should be empty.

When you update a maintenance window task that has options specified in TaskInvocationParameters, you must provide again all the TaskInvocationParameters values that you want to retain. The values you don't specify again are removed. For example, suppose that when you registered a Run Command task, you specified TaskInvocationParameters values for Comment, NotificationConfig, and OutputS3BucketName. If you update the maintenance window task and specify only a different OutputS3BucketName value, the values for Comment and NotificationConfig are removed.

name :: Maybe Text

The new task name to specify.

targets :: Maybe [Target]

The targets (either instances or tags) to modify. Instances are specified using the format Key=instanceids,Values=instanceID_1,instanceID_2. Tags are specified using the format Key=tag_name,Values=tag_value.

One or more targets must be specified for maintenance window Run Command-type tasks. Depending on the task, targets are optional for other maintenance window task types (Automation, Lambda, and Step Functions). For more information about running tasks that don't specify targets, see Registering maintenance window tasks without targets in the Amazon Web Services Systems Manager User Guide.

loggingInfo :: Maybe LoggingInfo

The new logging location in Amazon S3 to specify.

LoggingInfo has been deprecated. To specify an Amazon Simple Storage Service (Amazon S3) bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Amazon Web Services Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

description :: Maybe (Sensitive Text)

The new task description to specify.

maxConcurrency :: Maybe Text

The new MaxConcurrency value you want to specify. MaxConcurrency is the number of targets that are allowed to run this task in parallel.

For maintenance window tasks without a target specified, you can't supply a value for this option. Instead, the system inserts a placeholder value of 1, which may be reported in the response to this command. This value doesn't affect the running of your task and can be ignored.

windowId :: Text

The maintenance window ID that contains the task to modify.

windowTaskId :: Text

The task ID to modify.

newUpdateMaintenanceWindowTask

Arguments

:: Text
$sel:windowId:UpdateMaintenanceWindowTask'

-> Text
$sel:windowTaskId:UpdateMaintenanceWindowTask'

-> UpdateMaintenanceWindowTask

Create a value of UpdateMaintenanceWindowTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceRoleArn:UpdateMaintenanceWindowTask', updateMaintenanceWindowTask_serviceRoleArn - The Amazon Resource Name (ARN) of the IAM service role for Amazon Web Services Systems Manager to assume when running a maintenance window task. If you do not specify a service role ARN, Systems Manager uses your account's service-linked role. If no service-linked role for Systems Manager exists in your account, it is created when you run RegisterTaskWithMaintenanceWindow.

For more information, see the following topics in the in the /Amazon Web Services Systems Manager User Guide/:

Using service-linked roles for Systems Manager

Should I use a service-linked role or a custom service role to run maintenance window tasks?

$sel:replace:UpdateMaintenanceWindowTask', updateMaintenanceWindowTask_replace - If True, then all fields that are required by the RegisterTaskWithMaintenanceWindow operation are also required for this API request. Optional fields that aren't specified are set to null.

$sel:taskParameters:UpdateMaintenanceWindowTask', updateMaintenanceWindowTask_taskParameters - The parameters to modify.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

The map has the following format:

Key: string, between 1 and 255 characters

Value: an array of strings, each string is between 1 and 255 characters

$sel:priority:UpdateMaintenanceWindowTask', updateMaintenanceWindowTask_priority - The new task priority to specify. The lower the number, the higher the priority. Tasks that have the same priority are scheduled in parallel.

$sel:taskArn:UpdateMaintenanceWindowTask', updateMaintenanceWindowTask_taskArn - The task ARN to modify.

$sel:cutoffBehavior:UpdateMaintenanceWindowTask', updateMaintenanceWindowTask_cutoffBehavior - Indicates whether tasks should continue to run after the cutoff time specified in the maintenance windows is reached.

CONTINUE_TASK: When the cutoff time is reached, any tasks that are running continue. The default value.

CANCEL_TASK:

For Automation, Lambda, Step Functions tasks: When the cutoff time is reached, any task invocations that are already running continue, but no new task invocations are started.

For Run Command tasks: When the cutoff time is reached, the system sends a CancelCommand operation that attempts to cancel the command associated with the task. However, there is no guarantee that the command will be terminated and the underlying process stopped.

The status for tasks that are not completed is TIMED_OUT.

$sel:maxErrors:UpdateMaintenanceWindowTask', updateMaintenanceWindowTask_maxErrors - The new MaxErrors value to specify. MaxErrors is the maximum number of errors that are allowed before the task stops being scheduled.

For maintenance window tasks without a target specified, you can't supply a value for this option. Instead, the system inserts a placeholder value of 1, which may be reported in the response to this command. This value doesn't affect the running of your task and can be ignored.

$sel:taskInvocationParameters:UpdateMaintenanceWindowTask', updateMaintenanceWindowTask_taskInvocationParameters - The parameters that the task should use during execution. Populate only the fields that match the task type. All other fields should be empty.

When you update a maintenance window task that has options specified in TaskInvocationParameters, you must provide again all the TaskInvocationParameters values that you want to retain. The values you don't specify again are removed. For example, suppose that when you registered a Run Command task, you specified TaskInvocationParameters values for Comment, NotificationConfig, and OutputS3BucketName. If you update the maintenance window task and specify only a different OutputS3BucketName value, the values for Comment and NotificationConfig are removed.

$sel:name:UpdateMaintenanceWindowTask', updateMaintenanceWindowTask_name - The new task name to specify.

$sel:targets:UpdateMaintenanceWindowTask', updateMaintenanceWindowTask_targets - The targets (either instances or tags) to modify. Instances are specified using the format Key=instanceids,Values=instanceID_1,instanceID_2. Tags are specified using the format Key=tag_name,Values=tag_value.

One or more targets must be specified for maintenance window Run Command-type tasks. Depending on the task, targets are optional for other maintenance window task types (Automation, Lambda, and Step Functions). For more information about running tasks that don't specify targets, see Registering maintenance window tasks without targets in the Amazon Web Services Systems Manager User Guide.

$sel:loggingInfo:UpdateMaintenanceWindowTask', updateMaintenanceWindowTask_loggingInfo - The new logging location in Amazon S3 to specify.

LoggingInfo has been deprecated. To specify an Amazon Simple Storage Service (Amazon S3) bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Amazon Web Services Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

$sel:description:UpdateMaintenanceWindowTask', updateMaintenanceWindowTask_description - The new task description to specify.

$sel:maxConcurrency:UpdateMaintenanceWindowTask', updateMaintenanceWindowTask_maxConcurrency - The new MaxConcurrency value you want to specify. MaxConcurrency is the number of targets that are allowed to run this task in parallel.

For maintenance window tasks without a target specified, you can't supply a value for this option. Instead, the system inserts a placeholder value of 1, which may be reported in the response to this command. This value doesn't affect the running of your task and can be ignored.

$sel:windowId:UpdateMaintenanceWindowTask', updateMaintenanceWindowTask_windowId - The maintenance window ID that contains the task to modify.

$sel:windowTaskId:UpdateMaintenanceWindowTask', updateMaintenanceWindowTask_windowTaskId - The task ID to modify.

Request Lenses

updateMaintenanceWindowTask_serviceRoleArn :: Lens' UpdateMaintenanceWindowTask (Maybe Text)

The Amazon Resource Name (ARN) of the IAM service role for Amazon Web Services Systems Manager to assume when running a maintenance window task. If you do not specify a service role ARN, Systems Manager uses your account's service-linked role. If no service-linked role for Systems Manager exists in your account, it is created when you run RegisterTaskWithMaintenanceWindow.

For more information, see the following topics in the in the /Amazon Web Services Systems Manager User Guide/:

Using service-linked roles for Systems Manager

Should I use a service-linked role or a custom service role to run maintenance window tasks?

updateMaintenanceWindowTask_replace :: Lens' UpdateMaintenanceWindowTask (Maybe Bool)

If True, then all fields that are required by the RegisterTaskWithMaintenanceWindow operation are also required for this API request. Optional fields that aren't specified are set to null.

updateMaintenanceWindowTask_taskParameters :: Lens' UpdateMaintenanceWindowTask (Maybe (HashMap Text MaintenanceWindowTaskParameterValueExpression))

The parameters to modify.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

The map has the following format:

Key: string, between 1 and 255 characters

Value: an array of strings, each string is between 1 and 255 characters

updateMaintenanceWindowTask_priority :: Lens' UpdateMaintenanceWindowTask (Maybe Natural)

The new task priority to specify. The lower the number, the higher the priority. Tasks that have the same priority are scheduled in parallel.

updateMaintenanceWindowTask_taskArn :: Lens' UpdateMaintenanceWindowTask (Maybe Text)

The task ARN to modify.

updateMaintenanceWindowTask_cutoffBehavior :: Lens' UpdateMaintenanceWindowTask (Maybe MaintenanceWindowTaskCutoffBehavior)

Indicates whether tasks should continue to run after the cutoff time specified in the maintenance windows is reached.

CONTINUE_TASK: When the cutoff time is reached, any tasks that are running continue. The default value.

CANCEL_TASK:

For Automation, Lambda, Step Functions tasks: When the cutoff time is reached, any task invocations that are already running continue, but no new task invocations are started.

For Run Command tasks: When the cutoff time is reached, the system sends a CancelCommand operation that attempts to cancel the command associated with the task. However, there is no guarantee that the command will be terminated and the underlying process stopped.

The status for tasks that are not completed is TIMED_OUT.

updateMaintenanceWindowTask_maxErrors :: Lens' UpdateMaintenanceWindowTask (Maybe Text)

The new MaxErrors value to specify. MaxErrors is the maximum number of errors that are allowed before the task stops being scheduled.

For maintenance window tasks without a target specified, you can't supply a value for this option. Instead, the system inserts a placeholder value of 1, which may be reported in the response to this command. This value doesn't affect the running of your task and can be ignored.

updateMaintenanceWindowTask_taskInvocationParameters :: Lens' UpdateMaintenanceWindowTask (Maybe MaintenanceWindowTaskInvocationParameters)

The parameters that the task should use during execution. Populate only the fields that match the task type. All other fields should be empty.

When you update a maintenance window task that has options specified in TaskInvocationParameters, you must provide again all the TaskInvocationParameters values that you want to retain. The values you don't specify again are removed. For example, suppose that when you registered a Run Command task, you specified TaskInvocationParameters values for Comment, NotificationConfig, and OutputS3BucketName. If you update the maintenance window task and specify only a different OutputS3BucketName value, the values for Comment and NotificationConfig are removed.

updateMaintenanceWindowTask_name :: Lens' UpdateMaintenanceWindowTask (Maybe Text)

The new task name to specify.

updateMaintenanceWindowTask_targets :: Lens' UpdateMaintenanceWindowTask (Maybe [Target])

The targets (either instances or tags) to modify. Instances are specified using the format Key=instanceids,Values=instanceID_1,instanceID_2. Tags are specified using the format Key=tag_name,Values=tag_value.

One or more targets must be specified for maintenance window Run Command-type tasks. Depending on the task, targets are optional for other maintenance window task types (Automation, Lambda, and Step Functions). For more information about running tasks that don't specify targets, see Registering maintenance window tasks without targets in the Amazon Web Services Systems Manager User Guide.

updateMaintenanceWindowTask_loggingInfo :: Lens' UpdateMaintenanceWindowTask (Maybe LoggingInfo)

The new logging location in Amazon S3 to specify.

LoggingInfo has been deprecated. To specify an Amazon Simple Storage Service (Amazon S3) bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Amazon Web Services Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

updateMaintenanceWindowTask_description :: Lens' UpdateMaintenanceWindowTask (Maybe Text)

The new task description to specify.

updateMaintenanceWindowTask_maxConcurrency :: Lens' UpdateMaintenanceWindowTask (Maybe Text)

The new MaxConcurrency value you want to specify. MaxConcurrency is the number of targets that are allowed to run this task in parallel.

For maintenance window tasks without a target specified, you can't supply a value for this option. Instead, the system inserts a placeholder value of 1, which may be reported in the response to this command. This value doesn't affect the running of your task and can be ignored.

updateMaintenanceWindowTask_windowId :: Lens' UpdateMaintenanceWindowTask Text

The maintenance window ID that contains the task to modify.

updateMaintenanceWindowTask_windowTaskId :: Lens' UpdateMaintenanceWindowTask Text

The task ID to modify.

Destructuring the Response

data UpdateMaintenanceWindowTaskResponse

See: newUpdateMaintenanceWindowTaskResponse smart constructor.

Constructors

UpdateMaintenanceWindowTaskResponse'

Fields

serviceRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) service role to use to publish Amazon Simple Notification Service (Amazon SNS) notifications for maintenance window Run Command tasks.

windowTaskId :: Maybe Text

The task ID of the maintenance window that was updated.

taskParameters :: Maybe (Sensitive (HashMap Text (Sensitive MaintenanceWindowTaskParameterValueExpression)))

The updated parameter values.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

priority :: Maybe Natural

The updated priority value.

taskArn :: Maybe Text

The updated task ARN value.

cutoffBehavior :: Maybe MaintenanceWindowTaskCutoffBehavior

The specification for whether tasks should continue to run after the cutoff time specified in the maintenance windows is reached.

maxErrors :: Maybe Text

The updated MaxErrors value.

taskInvocationParameters :: Maybe MaintenanceWindowTaskInvocationParameters

The updated parameter values.

name :: Maybe Text

The updated task name.

targets :: Maybe [Target]

The updated target values.

loggingInfo :: Maybe LoggingInfo

The updated logging information in Amazon S3.

LoggingInfo has been deprecated. To specify an Amazon Simple Storage Service (Amazon S3) bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Amazon Web Services Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

description :: Maybe (Sensitive Text)

The updated task description.

maxConcurrency :: Maybe Text

The updated MaxConcurrency value.

windowId :: Maybe Text

The ID of the maintenance window that was updated.

httpStatus :: Int

The response's http status code.

newUpdateMaintenanceWindowTaskResponse

Arguments

:: Int
$sel:httpStatus:UpdateMaintenanceWindowTaskResponse'

-> UpdateMaintenanceWindowTaskResponse

Create a value of UpdateMaintenanceWindowTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceRoleArn:UpdateMaintenanceWindowTask', updateMaintenanceWindowTaskResponse_serviceRoleArn - The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) service role to use to publish Amazon Simple Notification Service (Amazon SNS) notifications for maintenance window Run Command tasks.

$sel:windowTaskId:UpdateMaintenanceWindowTask', updateMaintenanceWindowTaskResponse_windowTaskId - The task ID of the maintenance window that was updated.

$sel:taskParameters:UpdateMaintenanceWindowTask', updateMaintenanceWindowTaskResponse_taskParameters - The updated parameter values.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

$sel:priority:UpdateMaintenanceWindowTask', updateMaintenanceWindowTaskResponse_priority - The updated priority value.

$sel:taskArn:UpdateMaintenanceWindowTask', updateMaintenanceWindowTaskResponse_taskArn - The updated task ARN value.

$sel:cutoffBehavior:UpdateMaintenanceWindowTask', updateMaintenanceWindowTaskResponse_cutoffBehavior - The specification for whether tasks should continue to run after the cutoff time specified in the maintenance windows is reached.

$sel:maxErrors:UpdateMaintenanceWindowTask', updateMaintenanceWindowTaskResponse_maxErrors - The updated MaxErrors value.

$sel:taskInvocationParameters:UpdateMaintenanceWindowTask', updateMaintenanceWindowTaskResponse_taskInvocationParameters - The updated parameter values.

$sel:name:UpdateMaintenanceWindowTask', updateMaintenanceWindowTaskResponse_name - The updated task name.

$sel:targets:UpdateMaintenanceWindowTask', updateMaintenanceWindowTaskResponse_targets - The updated target values.

$sel:loggingInfo:UpdateMaintenanceWindowTask', updateMaintenanceWindowTaskResponse_loggingInfo - The updated logging information in Amazon S3.

LoggingInfo has been deprecated. To specify an Amazon Simple Storage Service (Amazon S3) bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Amazon Web Services Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

$sel:description:UpdateMaintenanceWindowTask', updateMaintenanceWindowTaskResponse_description - The updated task description.

$sel:maxConcurrency:UpdateMaintenanceWindowTask', updateMaintenanceWindowTaskResponse_maxConcurrency - The updated MaxConcurrency value.

$sel:windowId:UpdateMaintenanceWindowTask', updateMaintenanceWindowTaskResponse_windowId - The ID of the maintenance window that was updated.

$sel:httpStatus:UpdateMaintenanceWindowTaskResponse', updateMaintenanceWindowTaskResponse_httpStatus - The response's http status code.

Response Lenses

updateMaintenanceWindowTaskResponse_serviceRoleArn :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe Text)

The Amazon Resource Name (ARN) of the Identity and Access Management (IAM) service role to use to publish Amazon Simple Notification Service (Amazon SNS) notifications for maintenance window Run Command tasks.

updateMaintenanceWindowTaskResponse_windowTaskId :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe Text)

The task ID of the maintenance window that was updated.

updateMaintenanceWindowTaskResponse_taskParameters :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe (HashMap Text MaintenanceWindowTaskParameterValueExpression))

The updated parameter values.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

updateMaintenanceWindowTaskResponse_priority :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe Natural)

The updated priority value.

updateMaintenanceWindowTaskResponse_taskArn :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe Text)

The updated task ARN value.

updateMaintenanceWindowTaskResponse_cutoffBehavior :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe MaintenanceWindowTaskCutoffBehavior)

The specification for whether tasks should continue to run after the cutoff time specified in the maintenance windows is reached.

updateMaintenanceWindowTaskResponse_maxErrors :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe Text)

The updated MaxErrors value.

updateMaintenanceWindowTaskResponse_taskInvocationParameters :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe MaintenanceWindowTaskInvocationParameters)

The updated parameter values.

updateMaintenanceWindowTaskResponse_name :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe Text)

The updated task name.

updateMaintenanceWindowTaskResponse_targets :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe [Target])

The updated target values.

updateMaintenanceWindowTaskResponse_loggingInfo :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe LoggingInfo)

The updated logging information in Amazon S3.

LoggingInfo has been deprecated. To specify an Amazon Simple Storage Service (Amazon S3) bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Amazon Web Services Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

updateMaintenanceWindowTaskResponse_description :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe Text)

The updated task description.

updateMaintenanceWindowTaskResponse_maxConcurrency :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe Text)

The updated MaxConcurrency value.

updateMaintenanceWindowTaskResponse_windowId :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe Text)

The ID of the maintenance window that was updated.

updateMaintenanceWindowTaskResponse_httpStatus :: Lens' UpdateMaintenanceWindowTaskResponse Int

The response's http status code.