Network.AWS.SSM.UpdateOpsItem

Description

Edit or change an OpsItem. You must have permission in Identity and Access Management (IAM) to update an OpsItem. For more information, see Getting started with OpsCenter in the Amazon Web Services Systems Manager User Guide.

Operations engineers and IT professionals use Amazon Web Services Systems Manager OpsCenter to view, investigate, and remediate operational issues impacting the performance and health of their Amazon Web Services resources. For more information, see OpsCenter in the Amazon Web Services Systems Manager User Guide.

Creating a Request

data UpdateOpsItem

See: newUpdateOpsItem smart constructor.

Constructors

UpdateOpsItem'

Fields

status :: Maybe OpsItemStatus

The OpsItem status. Status can be Open, In Progress, or Resolved. For more information, see Editing OpsItem details in the Amazon Web Services Systems Manager User Guide.

plannedEndTime :: Maybe POSIX

The time specified in a change request for a runbook workflow to end. Currently supported only for the OpsItem type /aws/changerequest.

severity :: Maybe Text

Specify a new severity for an OpsItem.

actualStartTime :: Maybe POSIX

The time a runbook workflow started. Currently reported only for the OpsItem type /aws/changerequest.

category :: Maybe Text

Specify a new category for an OpsItem.

operationalData :: Maybe (HashMap Text OpsItemDataValue)

Add new keys or edit existing key-value pairs of the OperationalData map in the OpsItem object.

Operational data is custom data that provides useful reference details about the OpsItem. For example, you can specify log files, error strings, license keys, troubleshooting tips, or other relevant data. You enter operational data as key-value pairs. The key has a maximum length of 128 characters. The value has a maximum size of 20 KB.

Operational data keys can't begin with the following: amazon, aws, amzn, ssm, /amazon, /aws, /amzn, /ssm.

You can choose to make the data searchable by other users in the account or you can restrict search access. Searchable data means that all users with access to the OpsItem Overview page (as provided by the DescribeOpsItems API operation) can view and search on the specified data. Operational data that isn't searchable is only viewable by users who have access to the OpsItem (as provided by the GetOpsItem API operation).

Use the /aws/resources key in OperationalData to specify a related resource in the request. Use the /aws/automations key in OperationalData to associate an Automation runbook with the OpsItem. To view Amazon Web Services CLI example commands that use these keys, see Creating OpsItems manually in the Amazon Web Services Systems Manager User Guide.

title :: Maybe Text

A short heading that describes the nature of the OpsItem and the impacted resource.

priority :: Maybe Natural

The importance of this OpsItem in relation to other OpsItems in the system.

actualEndTime :: Maybe POSIX

The time a runbook workflow ended. Currently reported only for the OpsItem type /aws/changerequest.

plannedStartTime :: Maybe POSIX

The time specified in a change request for a runbook workflow to start. Currently supported only for the OpsItem type /aws/changerequest.

notifications :: Maybe [OpsItemNotification]

The Amazon Resource Name (ARN) of an SNS topic where notifications are sent when this OpsItem is edited or changed.

description :: Maybe Text

Update the information about the OpsItem. Provide enough information so that users reading this OpsItem for the first time understand the issue.

relatedOpsItems :: Maybe [RelatedOpsItem]

One or more OpsItems that share something in common with the current OpsItems. For example, related OpsItems can include OpsItems with similar error messages, impacted resources, or statuses for the impacted resource.

operationalDataToDelete :: Maybe [Text]

Keys that you want to remove from the OperationalData map.

opsItemId :: Text

The ID of the OpsItem.

newUpdateOpsItem

Arguments

:: Text
$sel:opsItemId:UpdateOpsItem'

-> UpdateOpsItem

Create a value of UpdateOpsItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UpdateOpsItem', updateOpsItem_status - The OpsItem status. Status can be Open, In Progress, or Resolved. For more information, see Editing OpsItem details in the Amazon Web Services Systems Manager User Guide.

$sel:plannedEndTime:UpdateOpsItem', updateOpsItem_plannedEndTime - The time specified in a change request for a runbook workflow to end. Currently supported only for the OpsItem type /aws/changerequest.

$sel:severity:UpdateOpsItem', updateOpsItem_severity - Specify a new severity for an OpsItem.

$sel:actualStartTime:UpdateOpsItem', updateOpsItem_actualStartTime - The time a runbook workflow started. Currently reported only for the OpsItem type /aws/changerequest.

$sel:category:UpdateOpsItem', updateOpsItem_category - Specify a new category for an OpsItem.

$sel:operationalData:UpdateOpsItem', updateOpsItem_operationalData - Add new keys or edit existing key-value pairs of the OperationalData map in the OpsItem object.

Operational data is custom data that provides useful reference details about the OpsItem. For example, you can specify log files, error strings, license keys, troubleshooting tips, or other relevant data. You enter operational data as key-value pairs. The key has a maximum length of 128 characters. The value has a maximum size of 20 KB.

Operational data keys can't begin with the following: amazon, aws, amzn, ssm, /amazon, /aws, /amzn, /ssm.

You can choose to make the data searchable by other users in the account or you can restrict search access. Searchable data means that all users with access to the OpsItem Overview page (as provided by the DescribeOpsItems API operation) can view and search on the specified data. Operational data that isn't searchable is only viewable by users who have access to the OpsItem (as provided by the GetOpsItem API operation).

Use the /aws/resources key in OperationalData to specify a related resource in the request. Use the /aws/automations key in OperationalData to associate an Automation runbook with the OpsItem. To view Amazon Web Services CLI example commands that use these keys, see Creating OpsItems manually in the Amazon Web Services Systems Manager User Guide.

$sel:title:UpdateOpsItem', updateOpsItem_title - A short heading that describes the nature of the OpsItem and the impacted resource.

$sel:priority:UpdateOpsItem', updateOpsItem_priority - The importance of this OpsItem in relation to other OpsItems in the system.

$sel:actualEndTime:UpdateOpsItem', updateOpsItem_actualEndTime - The time a runbook workflow ended. Currently reported only for the OpsItem type /aws/changerequest.

$sel:plannedStartTime:UpdateOpsItem', updateOpsItem_plannedStartTime - The time specified in a change request for a runbook workflow to start. Currently supported only for the OpsItem type /aws/changerequest.

$sel:notifications:UpdateOpsItem', updateOpsItem_notifications - The Amazon Resource Name (ARN) of an SNS topic where notifications are sent when this OpsItem is edited or changed.

$sel:description:UpdateOpsItem', updateOpsItem_description - Update the information about the OpsItem. Provide enough information so that users reading this OpsItem for the first time understand the issue.

$sel:relatedOpsItems:UpdateOpsItem', updateOpsItem_relatedOpsItems - One or more OpsItems that share something in common with the current OpsItems. For example, related OpsItems can include OpsItems with similar error messages, impacted resources, or statuses for the impacted resource.

$sel:operationalDataToDelete:UpdateOpsItem', updateOpsItem_operationalDataToDelete - Keys that you want to remove from the OperationalData map.

$sel:opsItemId:UpdateOpsItem', updateOpsItem_opsItemId - The ID of the OpsItem.

Request Lenses

updateOpsItem_status :: Lens' UpdateOpsItem (Maybe OpsItemStatus)

The OpsItem status. Status can be Open, In Progress, or Resolved. For more information, see Editing OpsItem details in the Amazon Web Services Systems Manager User Guide.

updateOpsItem_plannedEndTime :: Lens' UpdateOpsItem (Maybe UTCTime)

The time specified in a change request for a runbook workflow to end. Currently supported only for the OpsItem type /aws/changerequest.

updateOpsItem_severity :: Lens' UpdateOpsItem (Maybe Text)

Specify a new severity for an OpsItem.

updateOpsItem_actualStartTime :: Lens' UpdateOpsItem (Maybe UTCTime)

The time a runbook workflow started. Currently reported only for the OpsItem type /aws/changerequest.

updateOpsItem_category :: Lens' UpdateOpsItem (Maybe Text)

Specify a new category for an OpsItem.

updateOpsItem_operationalData :: Lens' UpdateOpsItem (Maybe (HashMap Text OpsItemDataValue))

Add new keys or edit existing key-value pairs of the OperationalData map in the OpsItem object.

Operational data is custom data that provides useful reference details about the OpsItem. For example, you can specify log files, error strings, license keys, troubleshooting tips, or other relevant data. You enter operational data as key-value pairs. The key has a maximum length of 128 characters. The value has a maximum size of 20 KB.

Operational data keys can't begin with the following: amazon, aws, amzn, ssm, /amazon, /aws, /amzn, /ssm.

You can choose to make the data searchable by other users in the account or you can restrict search access. Searchable data means that all users with access to the OpsItem Overview page (as provided by the DescribeOpsItems API operation) can view and search on the specified data. Operational data that isn't searchable is only viewable by users who have access to the OpsItem (as provided by the GetOpsItem API operation).

Use the /aws/resources key in OperationalData to specify a related resource in the request. Use the /aws/automations key in OperationalData to associate an Automation runbook with the OpsItem. To view Amazon Web Services CLI example commands that use these keys, see Creating OpsItems manually in the Amazon Web Services Systems Manager User Guide.

updateOpsItem_title :: Lens' UpdateOpsItem (Maybe Text)

A short heading that describes the nature of the OpsItem and the impacted resource.

updateOpsItem_priority :: Lens' UpdateOpsItem (Maybe Natural)

The importance of this OpsItem in relation to other OpsItems in the system.

updateOpsItem_actualEndTime :: Lens' UpdateOpsItem (Maybe UTCTime)

The time a runbook workflow ended. Currently reported only for the OpsItem type /aws/changerequest.

updateOpsItem_plannedStartTime :: Lens' UpdateOpsItem (Maybe UTCTime)

The time specified in a change request for a runbook workflow to start. Currently supported only for the OpsItem type /aws/changerequest.

updateOpsItem_notifications :: Lens' UpdateOpsItem (Maybe [OpsItemNotification])

The Amazon Resource Name (ARN) of an SNS topic where notifications are sent when this OpsItem is edited or changed.

updateOpsItem_description :: Lens' UpdateOpsItem (Maybe Text)

Update the information about the OpsItem. Provide enough information so that users reading this OpsItem for the first time understand the issue.

updateOpsItem_relatedOpsItems :: Lens' UpdateOpsItem (Maybe [RelatedOpsItem])

One or more OpsItems that share something in common with the current OpsItems. For example, related OpsItems can include OpsItems with similar error messages, impacted resources, or statuses for the impacted resource.

updateOpsItem_operationalDataToDelete :: Lens' UpdateOpsItem (Maybe [Text])

Keys that you want to remove from the OperationalData map.

updateOpsItem_opsItemId :: Lens' UpdateOpsItem Text

The ID of the OpsItem.

Destructuring the Response

data UpdateOpsItemResponse

See: newUpdateOpsItemResponse smart constructor.

Constructors

UpdateOpsItemResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateOpsItemResponse

Arguments

:: Int
$sel:httpStatus:UpdateOpsItemResponse'

-> UpdateOpsItemResponse

Create a value of UpdateOpsItemResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateOpsItemResponse', updateOpsItemResponse_httpStatus - The response's http status code.

Response Lenses

updateOpsItemResponse_httpStatus :: Lens' UpdateOpsItemResponse Int

The response's http status code.