Network.AWS.SSM.UpdateOpsMetadata

Description

Amazon Web Services Systems Manager calls this API operation when you edit OpsMetadata in Application Manager.

Creating a Request

data UpdateOpsMetadata

See: newUpdateOpsMetadata smart constructor.

Constructors

UpdateOpsMetadata'

Fields

metadataToUpdate :: Maybe (HashMap Text MetadataValue)

Metadata to add to an OpsMetadata object.

keysToDelete :: Maybe (NonEmpty Text)

The metadata keys to delete from the OpsMetadata object.

opsMetadataArn :: Text

The Amazon Resoure Name (ARN) of the OpsMetadata Object to update.

newUpdateOpsMetadata

Arguments

:: Text
$sel:opsMetadataArn:UpdateOpsMetadata'

-> UpdateOpsMetadata

Create a value of UpdateOpsMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metadataToUpdate:UpdateOpsMetadata', updateOpsMetadata_metadataToUpdate - Metadata to add to an OpsMetadata object.

$sel:keysToDelete:UpdateOpsMetadata', updateOpsMetadata_keysToDelete - The metadata keys to delete from the OpsMetadata object.

$sel:opsMetadataArn:UpdateOpsMetadata', updateOpsMetadata_opsMetadataArn - The Amazon Resoure Name (ARN) of the OpsMetadata Object to update.

Request Lenses

updateOpsMetadata_metadataToUpdate :: Lens' UpdateOpsMetadata (Maybe (HashMap Text MetadataValue))

Metadata to add to an OpsMetadata object.

updateOpsMetadata_keysToDelete :: Lens' UpdateOpsMetadata (Maybe (NonEmpty Text))

The metadata keys to delete from the OpsMetadata object.

updateOpsMetadata_opsMetadataArn :: Lens' UpdateOpsMetadata Text

The Amazon Resoure Name (ARN) of the OpsMetadata Object to update.

Destructuring the Response

data UpdateOpsMetadataResponse

See: newUpdateOpsMetadataResponse smart constructor.

Constructors

UpdateOpsMetadataResponse'

Fields

opsMetadataArn :: Maybe Text

The Amazon Resource Name (ARN) of the OpsMetadata Object that was updated.

httpStatus :: Int

The response's http status code.

newUpdateOpsMetadataResponse

Arguments

:: Int
$sel:httpStatus:UpdateOpsMetadataResponse'

-> UpdateOpsMetadataResponse

Create a value of UpdateOpsMetadataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:opsMetadataArn:UpdateOpsMetadata', updateOpsMetadataResponse_opsMetadataArn - The Amazon Resource Name (ARN) of the OpsMetadata Object that was updated.

$sel:httpStatus:UpdateOpsMetadataResponse', updateOpsMetadataResponse_httpStatus - The response's http status code.

Response Lenses

updateOpsMetadataResponse_opsMetadataArn :: Lens' UpdateOpsMetadataResponse (Maybe Text)

The Amazon Resource Name (ARN) of the OpsMetadata Object that was updated.

updateOpsMetadataResponse_httpStatus :: Lens' UpdateOpsMetadataResponse Int

The response's http status code.