Network.AWS.SSM.UpdatePatchBaseline

Description

Modifies an existing patch baseline. Fields not specified in the request are left unchanged.

For information about valid key-value pairs in PatchFilters for each supported operating system type, see PatchFilter.

Creating a Request

data UpdatePatchBaseline

See: newUpdatePatchBaseline smart constructor.

Constructors

UpdatePatchBaseline'

Fields

rejectedPatches :: Maybe [Text]

A list of explicitly rejected patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

sources :: Maybe [PatchSource]

Information about the patches to use to update the instances, including target operating systems and source repositories. Applies to Linux instances only.

approvedPatchesEnableNonSecurity :: Maybe Bool

Indicates whether the list of approved patches includes non-security updates that should be applied to the instances. The default value is false. Applies to Linux instances only.

approvedPatchesComplianceLevel :: Maybe PatchComplianceLevel

Assigns a new compliance severity level to an existing patch baseline.

name :: Maybe Text

The name of the patch baseline.

replace :: Maybe Bool

If True, then all fields that are required by the CreatePatchBaseline operation are also required for this API request. Optional fields that aren't specified are set to null.

description :: Maybe Text

A description of the patch baseline.

approvedPatches :: Maybe [Text]

A list of explicitly approved patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

rejectedPatchesAction :: Maybe PatchAction

The action for Patch Manager to take on patches included in the RejectedPackages list.

ALLOW_AS_DEPENDENCY : A package in the Rejected patches list is installed only if it is a dependency of another package. It is considered compliant with the patch baseline, and its status is reported as InstalledOther. This is the default action if no option is specified.

BLOCK : Packages in the RejectedPatches list, and packages that include them as dependencies, aren't installed under any circumstances. If a package was installed before it was added to the Rejected patches list, it is considered non-compliant with the patch baseline, and its status is reported as InstalledRejected.

globalFilters :: Maybe PatchFilterGroup

A set of global filters used to include patches in the baseline.

approvalRules :: Maybe PatchRuleGroup

A set of rules used to include patches in the baseline.

baselineId :: Text

The ID of the patch baseline to update.

newUpdatePatchBaseline

Arguments

:: Text
$sel:baselineId:UpdatePatchBaseline'

-> UpdatePatchBaseline

Create a value of UpdatePatchBaseline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rejectedPatches:UpdatePatchBaseline', updatePatchBaseline_rejectedPatches - A list of explicitly rejected patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

$sel:sources:UpdatePatchBaseline', updatePatchBaseline_sources - Information about the patches to use to update the instances, including target operating systems and source repositories. Applies to Linux instances only.

$sel:approvedPatchesEnableNonSecurity:UpdatePatchBaseline', updatePatchBaseline_approvedPatchesEnableNonSecurity - Indicates whether the list of approved patches includes non-security updates that should be applied to the instances. The default value is false. Applies to Linux instances only.

$sel:approvedPatchesComplianceLevel:UpdatePatchBaseline', updatePatchBaseline_approvedPatchesComplianceLevel - Assigns a new compliance severity level to an existing patch baseline.

$sel:name:UpdatePatchBaseline', updatePatchBaseline_name - The name of the patch baseline.

$sel:replace:UpdatePatchBaseline', updatePatchBaseline_replace - If True, then all fields that are required by the CreatePatchBaseline operation are also required for this API request. Optional fields that aren't specified are set to null.

$sel:description:UpdatePatchBaseline', updatePatchBaseline_description - A description of the patch baseline.

$sel:approvedPatches:UpdatePatchBaseline', updatePatchBaseline_approvedPatches - A list of explicitly approved patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

$sel:rejectedPatchesAction:UpdatePatchBaseline', updatePatchBaseline_rejectedPatchesAction - The action for Patch Manager to take on patches included in the RejectedPackages list.

ALLOW_AS_DEPENDENCY : A package in the Rejected patches list is installed only if it is a dependency of another package. It is considered compliant with the patch baseline, and its status is reported as InstalledOther. This is the default action if no option is specified.

BLOCK : Packages in the RejectedPatches list, and packages that include them as dependencies, aren't installed under any circumstances. If a package was installed before it was added to the Rejected patches list, it is considered non-compliant with the patch baseline, and its status is reported as InstalledRejected.

$sel:globalFilters:UpdatePatchBaseline', updatePatchBaseline_globalFilters - A set of global filters used to include patches in the baseline.

$sel:approvalRules:UpdatePatchBaseline', updatePatchBaseline_approvalRules - A set of rules used to include patches in the baseline.

$sel:baselineId:UpdatePatchBaseline', updatePatchBaseline_baselineId - The ID of the patch baseline to update.

Request Lenses

updatePatchBaseline_rejectedPatches :: Lens' UpdatePatchBaseline (Maybe [Text])

A list of explicitly rejected patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

updatePatchBaseline_sources :: Lens' UpdatePatchBaseline (Maybe [PatchSource])

Information about the patches to use to update the instances, including target operating systems and source repositories. Applies to Linux instances only.

updatePatchBaseline_approvedPatchesEnableNonSecurity :: Lens' UpdatePatchBaseline (Maybe Bool)

Indicates whether the list of approved patches includes non-security updates that should be applied to the instances. The default value is false. Applies to Linux instances only.

updatePatchBaseline_approvedPatchesComplianceLevel :: Lens' UpdatePatchBaseline (Maybe PatchComplianceLevel)

Assigns a new compliance severity level to an existing patch baseline.

updatePatchBaseline_name :: Lens' UpdatePatchBaseline (Maybe Text)

The name of the patch baseline.

updatePatchBaseline_replace :: Lens' UpdatePatchBaseline (Maybe Bool)

If True, then all fields that are required by the CreatePatchBaseline operation are also required for this API request. Optional fields that aren't specified are set to null.

updatePatchBaseline_description :: Lens' UpdatePatchBaseline (Maybe Text)

A description of the patch baseline.

updatePatchBaseline_approvedPatches :: Lens' UpdatePatchBaseline (Maybe [Text])

A list of explicitly approved patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

updatePatchBaseline_rejectedPatchesAction :: Lens' UpdatePatchBaseline (Maybe PatchAction)

The action for Patch Manager to take on patches included in the RejectedPackages list.

ALLOW_AS_DEPENDENCY : A package in the Rejected patches list is installed only if it is a dependency of another package. It is considered compliant with the patch baseline, and its status is reported as InstalledOther. This is the default action if no option is specified.

BLOCK : Packages in the RejectedPatches list, and packages that include them as dependencies, aren't installed under any circumstances. If a package was installed before it was added to the Rejected patches list, it is considered non-compliant with the patch baseline, and its status is reported as InstalledRejected.

updatePatchBaseline_globalFilters :: Lens' UpdatePatchBaseline (Maybe PatchFilterGroup)

A set of global filters used to include patches in the baseline.

updatePatchBaseline_approvalRules :: Lens' UpdatePatchBaseline (Maybe PatchRuleGroup)

A set of rules used to include patches in the baseline.

updatePatchBaseline_baselineId :: Lens' UpdatePatchBaseline Text

The ID of the patch baseline to update.

Destructuring the Response

data UpdatePatchBaselineResponse

See: newUpdatePatchBaselineResponse smart constructor.

Constructors

UpdatePatchBaselineResponse'

Fields

createdDate :: Maybe POSIX

The date when the patch baseline was created.

rejectedPatches :: Maybe [Text]

A list of explicitly rejected patches for the baseline.

baselineId :: Maybe Text

The ID of the deleted patch baseline.

sources :: Maybe [PatchSource]

Information about the patches to use to update the instances, including target operating systems and source repositories. Applies to Linux instances only.

approvedPatchesEnableNonSecurity :: Maybe Bool

Indicates whether the list of approved patches includes non-security updates that should be applied to the instances. The default value is false. Applies to Linux instances only.

approvedPatchesComplianceLevel :: Maybe PatchComplianceLevel

The compliance severity level assigned to the patch baseline after the update completed.

modifiedDate :: Maybe POSIX

The date when the patch baseline was last modified.

name :: Maybe Text

The name of the patch baseline.

description :: Maybe Text

A description of the patch baseline.

approvedPatches :: Maybe [Text]

A list of explicitly approved patches for the baseline.

rejectedPatchesAction :: Maybe PatchAction

The action specified to take on patches included in the RejectedPatches list. A patch can be allowed only if it is a dependency of another package, or blocked entirely along with packages that include it as a dependency.

operatingSystem :: Maybe OperatingSystem

The operating system rule used by the updated patch baseline.

globalFilters :: Maybe PatchFilterGroup

A set of global filters used to exclude patches from the baseline.

approvalRules :: Maybe PatchRuleGroup

A set of rules used to include patches in the baseline.

httpStatus :: Int

The response's http status code.

newUpdatePatchBaselineResponse

Arguments

:: Int
$sel:httpStatus:UpdatePatchBaselineResponse'

-> UpdatePatchBaselineResponse

Create a value of UpdatePatchBaselineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdDate:UpdatePatchBaselineResponse', updatePatchBaselineResponse_createdDate - The date when the patch baseline was created.

$sel:rejectedPatches:UpdatePatchBaseline', updatePatchBaselineResponse_rejectedPatches - A list of explicitly rejected patches for the baseline.

$sel:baselineId:UpdatePatchBaseline', updatePatchBaselineResponse_baselineId - The ID of the deleted patch baseline.

$sel:sources:UpdatePatchBaseline', updatePatchBaselineResponse_sources - Information about the patches to use to update the instances, including target operating systems and source repositories. Applies to Linux instances only.

$sel:approvedPatchesEnableNonSecurity:UpdatePatchBaseline', updatePatchBaselineResponse_approvedPatchesEnableNonSecurity - Indicates whether the list of approved patches includes non-security updates that should be applied to the instances. The default value is false. Applies to Linux instances only.

$sel:approvedPatchesComplianceLevel:UpdatePatchBaseline', updatePatchBaselineResponse_approvedPatchesComplianceLevel - The compliance severity level assigned to the patch baseline after the update completed.

$sel:modifiedDate:UpdatePatchBaselineResponse', updatePatchBaselineResponse_modifiedDate - The date when the patch baseline was last modified.

$sel:name:UpdatePatchBaseline', updatePatchBaselineResponse_name - The name of the patch baseline.

$sel:description:UpdatePatchBaseline', updatePatchBaselineResponse_description - A description of the patch baseline.

$sel:approvedPatches:UpdatePatchBaseline', updatePatchBaselineResponse_approvedPatches - A list of explicitly approved patches for the baseline.

$sel:rejectedPatchesAction:UpdatePatchBaseline', updatePatchBaselineResponse_rejectedPatchesAction - The action specified to take on patches included in the RejectedPatches list. A patch can be allowed only if it is a dependency of another package, or blocked entirely along with packages that include it as a dependency.

$sel:operatingSystem:UpdatePatchBaselineResponse', updatePatchBaselineResponse_operatingSystem - The operating system rule used by the updated patch baseline.

$sel:globalFilters:UpdatePatchBaseline', updatePatchBaselineResponse_globalFilters - A set of global filters used to exclude patches from the baseline.

$sel:approvalRules:UpdatePatchBaseline', updatePatchBaselineResponse_approvalRules - A set of rules used to include patches in the baseline.

$sel:httpStatus:UpdatePatchBaselineResponse', updatePatchBaselineResponse_httpStatus - The response's http status code.

Response Lenses

updatePatchBaselineResponse_createdDate :: Lens' UpdatePatchBaselineResponse (Maybe UTCTime)

The date when the patch baseline was created.

updatePatchBaselineResponse_rejectedPatches :: Lens' UpdatePatchBaselineResponse (Maybe [Text])

A list of explicitly rejected patches for the baseline.

updatePatchBaselineResponse_baselineId :: Lens' UpdatePatchBaselineResponse (Maybe Text)

The ID of the deleted patch baseline.

updatePatchBaselineResponse_sources :: Lens' UpdatePatchBaselineResponse (Maybe [PatchSource])

Information about the patches to use to update the instances, including target operating systems and source repositories. Applies to Linux instances only.

updatePatchBaselineResponse_approvedPatchesEnableNonSecurity :: Lens' UpdatePatchBaselineResponse (Maybe Bool)

Indicates whether the list of approved patches includes non-security updates that should be applied to the instances. The default value is false. Applies to Linux instances only.

updatePatchBaselineResponse_approvedPatchesComplianceLevel :: Lens' UpdatePatchBaselineResponse (Maybe PatchComplianceLevel)

The compliance severity level assigned to the patch baseline after the update completed.

updatePatchBaselineResponse_modifiedDate :: Lens' UpdatePatchBaselineResponse (Maybe UTCTime)

The date when the patch baseline was last modified.

updatePatchBaselineResponse_name :: Lens' UpdatePatchBaselineResponse (Maybe Text)

The name of the patch baseline.

updatePatchBaselineResponse_description :: Lens' UpdatePatchBaselineResponse (Maybe Text)

A description of the patch baseline.

updatePatchBaselineResponse_approvedPatches :: Lens' UpdatePatchBaselineResponse (Maybe [Text])

A list of explicitly approved patches for the baseline.

updatePatchBaselineResponse_rejectedPatchesAction :: Lens' UpdatePatchBaselineResponse (Maybe PatchAction)

The action specified to take on patches included in the RejectedPatches list. A patch can be allowed only if it is a dependency of another package, or blocked entirely along with packages that include it as a dependency.

updatePatchBaselineResponse_operatingSystem :: Lens' UpdatePatchBaselineResponse (Maybe OperatingSystem)

The operating system rule used by the updated patch baseline.

updatePatchBaselineResponse_globalFilters :: Lens' UpdatePatchBaselineResponse (Maybe PatchFilterGroup)

A set of global filters used to exclude patches from the baseline.

updatePatchBaselineResponse_approvalRules :: Lens' UpdatePatchBaselineResponse (Maybe PatchRuleGroup)

A set of rules used to include patches in the baseline.

updatePatchBaselineResponse_httpStatus :: Lens' UpdatePatchBaselineResponse Int

The response's http status code.