Network.AWS.SSM.UpdateResourceDataSync

Description

Update a resource data sync. After you create a resource data sync for a Region, you can't change the account options for that sync. For example, if you create a sync in the us-east-2 (Ohio) Region and you choose the Include only the current account option, you can't edit that sync later and choose the Include all accounts from my Organizations configuration option. Instead, you must delete the first resource data sync, and create a new one.

This API operation only supports a resource data sync that was created with a SyncFromSource SyncType.

Creating a Request

data UpdateResourceDataSync

See: newUpdateResourceDataSync smart constructor.

Constructors

UpdateResourceDataSync'

Fields

syncName :: Text

The name of the resource data sync you want to update.

syncType :: Text

The type of resource data sync. The supported SyncType is SyncFromSource.

syncSource :: ResourceDataSyncSource

Specify information about the data sources to synchronize.

newUpdateResourceDataSync

Arguments

:: Text
$sel:syncName:UpdateResourceDataSync'

-> Text
$sel:syncType:UpdateResourceDataSync'

-> ResourceDataSyncSource
$sel:syncSource:UpdateResourceDataSync'

-> UpdateResourceDataSync

Create a value of UpdateResourceDataSync with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:syncName:UpdateResourceDataSync', updateResourceDataSync_syncName - The name of the resource data sync you want to update.

$sel:syncType:UpdateResourceDataSync', updateResourceDataSync_syncType - The type of resource data sync. The supported SyncType is SyncFromSource.

$sel:syncSource:UpdateResourceDataSync', updateResourceDataSync_syncSource - Specify information about the data sources to synchronize.

Request Lenses

updateResourceDataSync_syncName :: Lens' UpdateResourceDataSync Text

The name of the resource data sync you want to update.

updateResourceDataSync_syncType :: Lens' UpdateResourceDataSync Text

The type of resource data sync. The supported SyncType is SyncFromSource.

updateResourceDataSync_syncSource :: Lens' UpdateResourceDataSync ResourceDataSyncSource

Specify information about the data sources to synchronize.

Destructuring the Response

data UpdateResourceDataSyncResponse

See: newUpdateResourceDataSyncResponse smart constructor.

Constructors

UpdateResourceDataSyncResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateResourceDataSyncResponse

Arguments

:: Int
$sel:httpStatus:UpdateResourceDataSyncResponse'

-> UpdateResourceDataSyncResponse

Create a value of UpdateResourceDataSyncResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateResourceDataSyncResponse', updateResourceDataSyncResponse_httpStatus - The response's http status code.

Response Lenses

updateResourceDataSyncResponse_httpStatus :: Lens' UpdateResourceDataSyncResponse Int

The response's http status code.