Network.AWS.SSMContacts.AcceptPage

Description

Used to acknowledge an engagement to a contact channel during an incident.

Creating a Request

data AcceptPage

See: newAcceptPage smart constructor.

Constructors

AcceptPage'

Fields

note :: Maybe Text

Information provided by the user when the user acknowledges the page.

contactChannelId :: Maybe Text

The ARN of the contact channel.

acceptCodeValidation :: Maybe AcceptCodeValidation

An optional field that Incident Manager uses to ENFORCE AcceptCode validation when acknowledging an page. Acknowledgement can occur by replying to a page, or when entering the AcceptCode in the console. Enforcing AcceptCode validation causes Incident Manager to verify that the code entered by the user matches the code sent by Incident Manager with the page.

Incident Manager can also IGNORE AcceptCode validation. Ignoring AcceptCode validation causes Incident Manager to accept any value entered for the AcceptCode.

pageId :: Text

The Amazon Resource Name (ARN) of the engagement to a contact channel.

acceptType :: AcceptType

The type indicates if the page was DELIVERED or READ.

acceptCode :: Text

The accept code is a 6-digit code used to acknowledge the page.

newAcceptPage

Arguments

:: Text
$sel:pageId:AcceptPage'

-> AcceptType
$sel:acceptType:AcceptPage'

-> Text
$sel:acceptCode:AcceptPage'

-> AcceptPage

Create a value of AcceptPage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:note:AcceptPage', acceptPage_note - Information provided by the user when the user acknowledges the page.

$sel:contactChannelId:AcceptPage', acceptPage_contactChannelId - The ARN of the contact channel.

$sel:acceptCodeValidation:AcceptPage', acceptPage_acceptCodeValidation - An optional field that Incident Manager uses to ENFORCE AcceptCode validation when acknowledging an page. Acknowledgement can occur by replying to a page, or when entering the AcceptCode in the console. Enforcing AcceptCode validation causes Incident Manager to verify that the code entered by the user matches the code sent by Incident Manager with the page.

Incident Manager can also IGNORE AcceptCode validation. Ignoring AcceptCode validation causes Incident Manager to accept any value entered for the AcceptCode.

$sel:pageId:AcceptPage', acceptPage_pageId - The Amazon Resource Name (ARN) of the engagement to a contact channel.

$sel:acceptType:AcceptPage', acceptPage_acceptType - The type indicates if the page was DELIVERED or READ.

$sel:acceptCode:AcceptPage', acceptPage_acceptCode - The accept code is a 6-digit code used to acknowledge the page.

Request Lenses

acceptPage_note :: Lens' AcceptPage (Maybe Text)

Information provided by the user when the user acknowledges the page.

acceptPage_contactChannelId :: Lens' AcceptPage (Maybe Text)

The ARN of the contact channel.

acceptPage_acceptCodeValidation :: Lens' AcceptPage (Maybe AcceptCodeValidation)

An optional field that Incident Manager uses to ENFORCE AcceptCode validation when acknowledging an page. Acknowledgement can occur by replying to a page, or when entering the AcceptCode in the console. Enforcing AcceptCode validation causes Incident Manager to verify that the code entered by the user matches the code sent by Incident Manager with the page.

Incident Manager can also IGNORE AcceptCode validation. Ignoring AcceptCode validation causes Incident Manager to accept any value entered for the AcceptCode.

acceptPage_pageId :: Lens' AcceptPage Text

The Amazon Resource Name (ARN) of the engagement to a contact channel.

acceptPage_acceptType :: Lens' AcceptPage AcceptType

The type indicates if the page was DELIVERED or READ.

acceptPage_acceptCode :: Lens' AcceptPage Text

The accept code is a 6-digit code used to acknowledge the page.

Destructuring the Response

data AcceptPageResponse

See: newAcceptPageResponse smart constructor.

Constructors

AcceptPageResponse'

Fields

httpStatus :: Int

The response's http status code.

newAcceptPageResponse

Arguments

:: Int
$sel:httpStatus:AcceptPageResponse'

-> AcceptPageResponse

Create a value of AcceptPageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AcceptPageResponse', acceptPageResponse_httpStatus - The response's http status code.

Response Lenses

acceptPageResponse_httpStatus :: Lens' AcceptPageResponse Int

The response's http status code.