Network.AWS.SSMContacts.ActivateContactChannel

Description

Activates a contact's contact channel. Incident Manager can't engage a contact until the contact channel has been activated.

Creating a Request

data ActivateContactChannel

See: newActivateContactChannel smart constructor.

Constructors

ActivateContactChannel'

Fields

contactChannelId :: Text

The Amazon Resource Name (ARN) of the contact channel.

activationCode :: Text

The code sent to the contact channel when it was created in the contact.

newActivateContactChannel

Arguments

:: Text
$sel:contactChannelId:ActivateContactChannel'

-> Text
$sel:activationCode:ActivateContactChannel'

-> ActivateContactChannel

Create a value of ActivateContactChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contactChannelId:ActivateContactChannel', activateContactChannel_contactChannelId - The Amazon Resource Name (ARN) of the contact channel.

$sel:activationCode:ActivateContactChannel', activateContactChannel_activationCode - The code sent to the contact channel when it was created in the contact.

Request Lenses

activateContactChannel_contactChannelId :: Lens' ActivateContactChannel Text

The Amazon Resource Name (ARN) of the contact channel.

activateContactChannel_activationCode :: Lens' ActivateContactChannel Text

The code sent to the contact channel when it was created in the contact.

Destructuring the Response

data ActivateContactChannelResponse

See: newActivateContactChannelResponse smart constructor.

Constructors

ActivateContactChannelResponse'

Fields

httpStatus :: Int

The response's http status code.

newActivateContactChannelResponse

Arguments

:: Int
$sel:httpStatus:ActivateContactChannelResponse'

-> ActivateContactChannelResponse

Create a value of ActivateContactChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ActivateContactChannelResponse', activateContactChannelResponse_httpStatus - The response's http status code.

Response Lenses

activateContactChannelResponse_httpStatus :: Lens' ActivateContactChannelResponse Int

The response's http status code.