Network.AWS.SSMContacts.DescribeEngagement

Description

Incident Manager uses engagements to engage contacts and escalation plans during an incident. Use this command to describe the engagement that occurred during an incident.

Creating a Request

data DescribeEngagement

See: newDescribeEngagement smart constructor.

Constructors

DescribeEngagement'

Fields

engagementId :: Text

The Amazon Resource Name (ARN) of the engagement you want the details of.

newDescribeEngagement

Arguments

:: Text
$sel:engagementId:DescribeEngagement'

-> DescribeEngagement

Create a value of DescribeEngagement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engagementId:DescribeEngagement', describeEngagement_engagementId - The Amazon Resource Name (ARN) of the engagement you want the details of.

Request Lenses

describeEngagement_engagementId :: Lens' DescribeEngagement Text

The Amazon Resource Name (ARN) of the engagement you want the details of.

Destructuring the Response

data DescribeEngagementResponse

See: newDescribeEngagementResponse smart constructor.

Constructors

DescribeEngagementResponse'

Fields

publicSubject :: Maybe Text

The insecure subject of the message that was sent to the contact. Use this field for engagements to SMS.

startTime :: Maybe POSIX

The time that the engagement started.

publicContent :: Maybe Text

The insecure content of the message that was sent to the contact. Use this field for engagements to SMS.

stopTime :: Maybe POSIX

The time that the engagement ended.

incidentId :: Maybe Text

The ARN of the incident in which the engagement occurred.

httpStatus :: Int

The response's http status code.

contactArn :: Text

The ARN of the escalation plan or contacts involved in the engagement.

engagementArn :: Text

The ARN of the engagement.

sender :: Text

The user that started the engagement.

subject :: Text

The secure subject of the message that was sent to the contact. Use this field for engagements to VOICE and EMAIL.

content :: Text

The secure content of the message that was sent to the contact. Use this field for engagements to VOICE and EMAIL.

newDescribeEngagementResponse

Arguments

:: Int
$sel:httpStatus:DescribeEngagementResponse'

-> Text
$sel:contactArn:DescribeEngagementResponse'

-> Text
$sel:engagementArn:DescribeEngagementResponse'

-> Text
$sel:sender:DescribeEngagementResponse'

-> Text
$sel:subject:DescribeEngagementResponse'

-> Text
$sel:content:DescribeEngagementResponse'

-> DescribeEngagementResponse

Create a value of DescribeEngagementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:publicSubject:DescribeEngagementResponse', describeEngagementResponse_publicSubject - The insecure subject of the message that was sent to the contact. Use this field for engagements to SMS.

$sel:startTime:DescribeEngagementResponse', describeEngagementResponse_startTime - The time that the engagement started.

$sel:publicContent:DescribeEngagementResponse', describeEngagementResponse_publicContent - The insecure content of the message that was sent to the contact. Use this field for engagements to SMS.

$sel:stopTime:DescribeEngagementResponse', describeEngagementResponse_stopTime - The time that the engagement ended.

$sel:incidentId:DescribeEngagementResponse', describeEngagementResponse_incidentId - The ARN of the incident in which the engagement occurred.

$sel:httpStatus:DescribeEngagementResponse', describeEngagementResponse_httpStatus - The response's http status code.

$sel:contactArn:DescribeEngagementResponse', describeEngagementResponse_contactArn - The ARN of the escalation plan or contacts involved in the engagement.

$sel:engagementArn:DescribeEngagementResponse', describeEngagementResponse_engagementArn - The ARN of the engagement.

$sel:sender:DescribeEngagementResponse', describeEngagementResponse_sender - The user that started the engagement.

$sel:subject:DescribeEngagementResponse', describeEngagementResponse_subject - The secure subject of the message that was sent to the contact. Use this field for engagements to VOICE and EMAIL.

$sel:content:DescribeEngagementResponse', describeEngagementResponse_content - The secure content of the message that was sent to the contact. Use this field for engagements to VOICE and EMAIL.

Response Lenses

describeEngagementResponse_publicSubject :: Lens' DescribeEngagementResponse (Maybe Text)

The insecure subject of the message that was sent to the contact. Use this field for engagements to SMS.

describeEngagementResponse_startTime :: Lens' DescribeEngagementResponse (Maybe UTCTime)

The time that the engagement started.

describeEngagementResponse_publicContent :: Lens' DescribeEngagementResponse (Maybe Text)

The insecure content of the message that was sent to the contact. Use this field for engagements to SMS.

describeEngagementResponse_stopTime :: Lens' DescribeEngagementResponse (Maybe UTCTime)

The time that the engagement ended.

describeEngagementResponse_incidentId :: Lens' DescribeEngagementResponse (Maybe Text)

The ARN of the incident in which the engagement occurred.

describeEngagementResponse_httpStatus :: Lens' DescribeEngagementResponse Int

The response's http status code.

describeEngagementResponse_contactArn :: Lens' DescribeEngagementResponse Text

The ARN of the escalation plan or contacts involved in the engagement.

describeEngagementResponse_engagementArn :: Lens' DescribeEngagementResponse Text

The ARN of the engagement.

describeEngagementResponse_sender :: Lens' DescribeEngagementResponse Text

The user that started the engagement.

describeEngagementResponse_subject :: Lens' DescribeEngagementResponse Text

The secure subject of the message that was sent to the contact. Use this field for engagements to VOICE and EMAIL.

describeEngagementResponse_content :: Lens' DescribeEngagementResponse Text

The secure content of the message that was sent to the contact. Use this field for engagements to VOICE and EMAIL.