Network.AWS.SSMContacts.DescribePage

Description

Lists details of the engagement to a contact channel.

Creating a Request

data DescribePage

See: newDescribePage smart constructor.

Constructors

DescribePage'

Fields

pageId :: Text

The ID of the engagement to a contact channel.

newDescribePage

Arguments

:: Text
$sel:pageId:DescribePage'

-> DescribePage

Create a value of DescribePage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageId:DescribePage', describePage_pageId - The ID of the engagement to a contact channel.

Request Lenses

describePage_pageId :: Lens' DescribePage Text

The ID of the engagement to a contact channel.

Destructuring the Response

data DescribePageResponse

See: newDescribePageResponse smart constructor.

Constructors

DescribePageResponse'

Fields

readTime :: Maybe POSIX

The time that the contact channel acknowledged the engagement.

publicSubject :: Maybe Text

The insecure subject of the message that was sent to the contact. Use this field for engagements to SMS.

publicContent :: Maybe Text

The insecure content of the message that was sent to the contact. Use this field for engagements to SMS.

deliveryTime :: Maybe POSIX

The time that the contact channel received the engagement.

incidentId :: Maybe Text

The ARN of the incident that engaged the contact channel.

sentTime :: Maybe POSIX

The time the engagement was sent to the contact channel.

httpStatus :: Int

The response's http status code.

pageArn :: Text

The Amazon Resource Name (ARN) of the engagement to a contact channel.

engagementArn :: Text

The ARN of the engagement that engaged the contact channel.

contactArn :: Text

The ARN of the contact that was engaged.

sender :: Text

The user that started the engagement.

subject :: Text

The secure subject of the message that was sent to the contact. Use this field for engagements to VOICE and EMAIL.

content :: Text

The secure content of the message that was sent to the contact. Use this field for engagements to VOICE and EMAIL.

newDescribePageResponse

Arguments

:: Int
$sel:httpStatus:DescribePageResponse'

-> Text
$sel:pageArn:DescribePageResponse'

-> Text
$sel:engagementArn:DescribePageResponse'

-> Text
$sel:contactArn:DescribePageResponse'

-> Text
$sel:sender:DescribePageResponse'

-> Text
$sel:subject:DescribePageResponse'

-> Text
$sel:content:DescribePageResponse'

-> DescribePageResponse

Create a value of DescribePageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:readTime:DescribePageResponse', describePageResponse_readTime - The time that the contact channel acknowledged the engagement.

$sel:publicSubject:DescribePageResponse', describePageResponse_publicSubject - The insecure subject of the message that was sent to the contact. Use this field for engagements to SMS.

$sel:publicContent:DescribePageResponse', describePageResponse_publicContent - The insecure content of the message that was sent to the contact. Use this field for engagements to SMS.

$sel:deliveryTime:DescribePageResponse', describePageResponse_deliveryTime - The time that the contact channel received the engagement.

$sel:incidentId:DescribePageResponse', describePageResponse_incidentId - The ARN of the incident that engaged the contact channel.

$sel:sentTime:DescribePageResponse', describePageResponse_sentTime - The time the engagement was sent to the contact channel.

$sel:httpStatus:DescribePageResponse', describePageResponse_httpStatus - The response's http status code.

$sel:pageArn:DescribePageResponse', describePageResponse_pageArn - The Amazon Resource Name (ARN) of the engagement to a contact channel.

$sel:engagementArn:DescribePageResponse', describePageResponse_engagementArn - The ARN of the engagement that engaged the contact channel.

$sel:contactArn:DescribePageResponse', describePageResponse_contactArn - The ARN of the contact that was engaged.

$sel:sender:DescribePageResponse', describePageResponse_sender - The user that started the engagement.

$sel:subject:DescribePageResponse', describePageResponse_subject - The secure subject of the message that was sent to the contact. Use this field for engagements to VOICE and EMAIL.

$sel:content:DescribePageResponse', describePageResponse_content - The secure content of the message that was sent to the contact. Use this field for engagements to VOICE and EMAIL.

Response Lenses

describePageResponse_readTime :: Lens' DescribePageResponse (Maybe UTCTime)

The time that the contact channel acknowledged the engagement.

describePageResponse_publicSubject :: Lens' DescribePageResponse (Maybe Text)

The insecure subject of the message that was sent to the contact. Use this field for engagements to SMS.

describePageResponse_publicContent :: Lens' DescribePageResponse (Maybe Text)

The insecure content of the message that was sent to the contact. Use this field for engagements to SMS.

describePageResponse_deliveryTime :: Lens' DescribePageResponse (Maybe UTCTime)

The time that the contact channel received the engagement.

describePageResponse_incidentId :: Lens' DescribePageResponse (Maybe Text)

The ARN of the incident that engaged the contact channel.

describePageResponse_sentTime :: Lens' DescribePageResponse (Maybe UTCTime)

The time the engagement was sent to the contact channel.

describePageResponse_httpStatus :: Lens' DescribePageResponse Int

The response's http status code.

describePageResponse_pageArn :: Lens' DescribePageResponse Text

The Amazon Resource Name (ARN) of the engagement to a contact channel.

describePageResponse_engagementArn :: Lens' DescribePageResponse Text

The ARN of the engagement that engaged the contact channel.

describePageResponse_contactArn :: Lens' DescribePageResponse Text

The ARN of the contact that was engaged.

describePageResponse_sender :: Lens' DescribePageResponse Text

The user that started the engagement.

describePageResponse_subject :: Lens' DescribePageResponse Text

The secure subject of the message that was sent to the contact. Use this field for engagements to VOICE and EMAIL.

describePageResponse_content :: Lens' DescribePageResponse Text

The secure content of the message that was sent to the contact. Use this field for engagements to VOICE and EMAIL.