Network.AWS.SSMContacts.GetContactChannel

Description

List details about a specific contact channel.

Creating a Request

data GetContactChannel

See: newGetContactChannel smart constructor.

Constructors

GetContactChannel'

Fields

contactChannelId :: Text

The Amazon Resource Name (ARN) of the contact channel you want information about.

newGetContactChannel

Arguments

:: Text
$sel:contactChannelId:GetContactChannel'

-> GetContactChannel

Create a value of GetContactChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contactChannelId:GetContactChannel', getContactChannel_contactChannelId - The Amazon Resource Name (ARN) of the contact channel you want information about.

Request Lenses

getContactChannel_contactChannelId :: Lens' GetContactChannel Text

The Amazon Resource Name (ARN) of the contact channel you want information about.

Destructuring the Response

data GetContactChannelResponse

See: newGetContactChannelResponse smart constructor.

Constructors

GetContactChannelResponse'

Fields

activationStatus :: Maybe ActivationStatus

A Boolean value indicating if the contact channel has been activated or not.

httpStatus :: Int

The response's http status code.

contactArn :: Text

The ARN of the contact that the channel belongs to.

contactChannelArn :: Text

The ARN of the contact channel.

name :: Text

The name of the contact channel

type' :: ChannelType

The type of contact channel. The type is SMS, VOICE, or EMAIL.

deliveryAddress :: ContactChannelAddress

The details that Incident Manager uses when trying to engage the contact channel.

newGetContactChannelResponse

Arguments

:: Int
$sel:httpStatus:GetContactChannelResponse'

-> Text
$sel:contactArn:GetContactChannelResponse'

-> Text
$sel:contactChannelArn:GetContactChannelResponse'

-> Text
$sel:name:GetContactChannelResponse'

-> ChannelType
$sel:type':GetContactChannelResponse'

-> ContactChannelAddress
$sel:deliveryAddress:GetContactChannelResponse'

-> GetContactChannelResponse

Create a value of GetContactChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:activationStatus:GetContactChannelResponse', getContactChannelResponse_activationStatus - A Boolean value indicating if the contact channel has been activated or not.

$sel:httpStatus:GetContactChannelResponse', getContactChannelResponse_httpStatus - The response's http status code.

$sel:contactArn:GetContactChannelResponse', getContactChannelResponse_contactArn - The ARN of the contact that the channel belongs to.

$sel:contactChannelArn:GetContactChannelResponse', getContactChannelResponse_contactChannelArn - The ARN of the contact channel.

$sel:name:GetContactChannelResponse', getContactChannelResponse_name - The name of the contact channel

$sel:type':GetContactChannelResponse', getContactChannelResponse_type - The type of contact channel. The type is SMS, VOICE, or EMAIL.

$sel:deliveryAddress:GetContactChannelResponse', getContactChannelResponse_deliveryAddress - The details that Incident Manager uses when trying to engage the contact channel.

Response Lenses

getContactChannelResponse_activationStatus :: Lens' GetContactChannelResponse (Maybe ActivationStatus)

A Boolean value indicating if the contact channel has been activated or not.

getContactChannelResponse_httpStatus :: Lens' GetContactChannelResponse Int

The response's http status code.

getContactChannelResponse_contactArn :: Lens' GetContactChannelResponse Text

The ARN of the contact that the channel belongs to.

getContactChannelResponse_contactChannelArn :: Lens' GetContactChannelResponse Text

The ARN of the contact channel.

getContactChannelResponse_name :: Lens' GetContactChannelResponse Text

The name of the contact channel

getContactChannelResponse_type :: Lens' GetContactChannelResponse ChannelType

The type of contact channel. The type is SMS, VOICE, or EMAIL.

getContactChannelResponse_deliveryAddress :: Lens' GetContactChannelResponse ContactChannelAddress

The details that Incident Manager uses when trying to engage the contact channel.