Network.AWS.SSMContacts.StartEngagement

Description

Starts an engagement to a contact or escalation plan. The engagement engages each contact specified in the incident.

Creating a Request

data StartEngagement

See: newStartEngagement smart constructor.

Constructors

StartEngagement'

Fields

idempotencyToken :: Maybe Text

A token ensuring that the operation is called only once with the specified details.

publicSubject :: Maybe Text

The insecure subject of the message that was sent to the contact. Use this field for engagements to SMS.

publicContent :: Maybe Text

The insecure content of the message that was sent to the contact. Use this field for engagements to SMS.

incidentId :: Maybe Text

The ARN of the incident that the engagement is part of.

contactId :: Text

The Amazon Resource Name (ARN) of the contact being engaged.

sender :: Text

The user that started the engagement.

subject :: Text

The secure subject of the message that was sent to the contact. Use this field for engagements to VOICE or EMAIL.

content :: Text

The secure content of the message that was sent to the contact. Use this field for engagements to VOICE or EMAIL.

newStartEngagement

Arguments

:: Text
$sel:contactId:StartEngagement'

-> Text
$sel:sender:StartEngagement'

-> Text
$sel:subject:StartEngagement'

-> Text
$sel:content:StartEngagement'

-> StartEngagement

Create a value of StartEngagement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:idempotencyToken:StartEngagement', startEngagement_idempotencyToken - A token ensuring that the operation is called only once with the specified details.

$sel:publicSubject:StartEngagement', startEngagement_publicSubject - The insecure subject of the message that was sent to the contact. Use this field for engagements to SMS.

$sel:publicContent:StartEngagement', startEngagement_publicContent - The insecure content of the message that was sent to the contact. Use this field for engagements to SMS.

$sel:incidentId:StartEngagement', startEngagement_incidentId - The ARN of the incident that the engagement is part of.

$sel:contactId:StartEngagement', startEngagement_contactId - The Amazon Resource Name (ARN) of the contact being engaged.

$sel:sender:StartEngagement', startEngagement_sender - The user that started the engagement.

$sel:subject:StartEngagement', startEngagement_subject - The secure subject of the message that was sent to the contact. Use this field for engagements to VOICE or EMAIL.

$sel:content:StartEngagement', startEngagement_content - The secure content of the message that was sent to the contact. Use this field for engagements to VOICE or EMAIL.

Request Lenses

startEngagement_idempotencyToken :: Lens' StartEngagement (Maybe Text)

A token ensuring that the operation is called only once with the specified details.

startEngagement_publicSubject :: Lens' StartEngagement (Maybe Text)

The insecure subject of the message that was sent to the contact. Use this field for engagements to SMS.

startEngagement_publicContent :: Lens' StartEngagement (Maybe Text)

The insecure content of the message that was sent to the contact. Use this field for engagements to SMS.

startEngagement_incidentId :: Lens' StartEngagement (Maybe Text)

The ARN of the incident that the engagement is part of.

startEngagement_contactId :: Lens' StartEngagement Text

The Amazon Resource Name (ARN) of the contact being engaged.

startEngagement_sender :: Lens' StartEngagement Text

The user that started the engagement.

startEngagement_subject :: Lens' StartEngagement Text

The secure subject of the message that was sent to the contact. Use this field for engagements to VOICE or EMAIL.

startEngagement_content :: Lens' StartEngagement Text

The secure content of the message that was sent to the contact. Use this field for engagements to VOICE or EMAIL.

Destructuring the Response

data StartEngagementResponse

See: newStartEngagementResponse smart constructor.

Constructors

StartEngagementResponse'

Fields

httpStatus :: Int

The response's http status code.

engagementArn :: Text

The ARN of the engagement.

newStartEngagementResponse

Arguments

:: Int
$sel:httpStatus:StartEngagementResponse'

-> Text
$sel:engagementArn:StartEngagementResponse'

-> StartEngagementResponse

Create a value of StartEngagementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StartEngagementResponse', startEngagementResponse_httpStatus - The response's http status code.

$sel:engagementArn:StartEngagementResponse', startEngagementResponse_engagementArn - The ARN of the engagement.

Response Lenses

startEngagementResponse_httpStatus :: Lens' StartEngagementResponse Int

The response's http status code.

startEngagementResponse_engagementArn :: Lens' StartEngagementResponse Text

The ARN of the engagement.