Network.AWS.SSMContacts.Types

Description

Service Configuration

defaultService :: Service

API version 2021-05-03 of the Amazon Systems Manager Incident Manager Contacts SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by an Amazon Web Services service.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You don't have sufficient access to perform this operation.

_DataEncryptionException :: AsError a => Getting (First ServiceError) a ServiceError

The operation failed to due an encryption key error.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Updating or deleting a resource causes an inconsistent state.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Request would cause a service quota to be exceeded.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

Unexpected error occurred while processing the request.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Request references a resource that doesn't exist.

AcceptCodeValidation

newtype AcceptCodeValidation

Constructors

AcceptCodeValidation'

Fields

fromAcceptCodeValidation :: Text

Bundled Patterns

pattern AcceptCodeValidation_ENFORCE :: AcceptCodeValidation

pattern AcceptCodeValidation_IGNORE :: AcceptCodeValidation

AcceptType

newtype AcceptType

Constructors

AcceptType'

Fields

fromAcceptType :: Text

Bundled Patterns

pattern AcceptType_DELIVERED :: AcceptType

pattern AcceptType_READ :: AcceptType

ActivationStatus

newtype ActivationStatus

Constructors

ActivationStatus'

Fields

fromActivationStatus :: Text

Bundled Patterns

pattern ActivationStatus_ACTIVATED :: ActivationStatus

pattern ActivationStatus_NOT_ACTIVATED :: ActivationStatus

ChannelType

newtype ChannelType

Constructors

ChannelType'

Fields

fromChannelType :: Text

Bundled Patterns

pattern ChannelType_EMAIL :: ChannelType

pattern ChannelType_SMS :: ChannelType

pattern ChannelType_VOICE :: ChannelType

ContactType

newtype ContactType

Constructors

ContactType'

Fields

fromContactType :: Text

Bundled Patterns

pattern ContactType_ESCALATION :: ContactType

pattern ContactType_PERSONAL :: ContactType

ReceiptType

newtype ReceiptType

Constructors

ReceiptType'

Fields

fromReceiptType :: Text

Bundled Patterns

pattern ReceiptType_DELIVERED :: ReceiptType

pattern ReceiptType_ERROR :: ReceiptType

pattern ReceiptType_READ :: ReceiptType

pattern ReceiptType_SENT :: ReceiptType

pattern ReceiptType_STOP :: ReceiptType

ChannelTargetInfo

data ChannelTargetInfo

Information about the contact channel that Incident Manager uses to engage the contact.

See: newChannelTargetInfo smart constructor.

Constructors

ChannelTargetInfo'

Fields

retryIntervalInMinutes :: Maybe Natural

The number of minutes to wait to retry sending engagement in the case the engagement initially fails.

contactChannelId :: Text

The Amazon Resource Name (ARN) of the contact channel.

newChannelTargetInfo

Arguments

:: Text
$sel:contactChannelId:ChannelTargetInfo'

-> ChannelTargetInfo

Create a value of ChannelTargetInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retryIntervalInMinutes:ChannelTargetInfo', channelTargetInfo_retryIntervalInMinutes - The number of minutes to wait to retry sending engagement in the case the engagement initially fails.

$sel:contactChannelId:ChannelTargetInfo', channelTargetInfo_contactChannelId - The Amazon Resource Name (ARN) of the contact channel.

channelTargetInfo_retryIntervalInMinutes :: Lens' ChannelTargetInfo (Maybe Natural)

The number of minutes to wait to retry sending engagement in the case the engagement initially fails.

channelTargetInfo_contactChannelId :: Lens' ChannelTargetInfo Text

The Amazon Resource Name (ARN) of the contact channel.

Contact

data Contact

A personal contact or escalation plan that Incident Manager engages during an incident.

See: newContact smart constructor.

Constructors

Contact'

Fields

displayName :: Maybe Text

The full name of the contact or escalation plan.

contactArn :: Text

The Amazon Resource Name (ARN) of the contact or escalation plan.

alias :: Text

The unique and identifiable alias of the contact or escalation plan.

type' :: ContactType

Refers to the type of contact. A single contact is type PERSONAL and an escalation plan is type ESCALATION.

newContact

Arguments

:: Text
$sel:contactArn:Contact'

-> Text
$sel:alias:Contact'

-> ContactType
$sel:type':Contact'

-> Contact

Create a value of Contact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:displayName:Contact', contact_displayName - The full name of the contact or escalation plan.

$sel:contactArn:Contact', contact_contactArn - The Amazon Resource Name (ARN) of the contact or escalation plan.

$sel:alias:Contact', contact_alias - The unique and identifiable alias of the contact or escalation plan.

$sel:type':Contact', contact_type - Refers to the type of contact. A single contact is type PERSONAL and an escalation plan is type ESCALATION.

contact_displayName :: Lens' Contact (Maybe Text)

The full name of the contact or escalation plan.

contact_contactArn :: Lens' Contact Text

The Amazon Resource Name (ARN) of the contact or escalation plan.

contact_alias :: Lens' Contact Text

The unique and identifiable alias of the contact or escalation plan.

contact_type :: Lens' Contact ContactType

Refers to the type of contact. A single contact is type PERSONAL and an escalation plan is type ESCALATION.

ContactChannel

data ContactChannel

The method that Incident Manager uses to engage a contact.

See: newContactChannel smart constructor.

Constructors

ContactChannel'

Fields

type' :: Maybe ChannelType

The type of the contact channel. Incident Manager supports three contact methods:

SMS

VOICE

EMAIL

contactChannelArn :: Text

The Amazon Resource Name (ARN) of the contact channel.

contactArn :: Text

The ARN of the contact that contains the contact channel.

name :: Text

The name of the contact channel.

deliveryAddress :: ContactChannelAddress

The details that Incident Manager uses when trying to engage the contact channel.

activationStatus :: ActivationStatus

A Boolean value describing if the contact channel has been activated or not. If the contact channel isn't activated, Incident Manager can't engage the contact through it.

newContactChannel

Arguments

:: Text
$sel:contactChannelArn:ContactChannel'

-> Text
$sel:contactArn:ContactChannel'

-> Text
$sel:name:ContactChannel'

-> ContactChannelAddress
$sel:deliveryAddress:ContactChannel'

-> ActivationStatus
$sel:activationStatus:ContactChannel'

-> ContactChannel

Create a value of ContactChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':ContactChannel', contactChannel_type - The type of the contact channel. Incident Manager supports three contact methods:

SMS

VOICE

EMAIL

$sel:contactChannelArn:ContactChannel', contactChannel_contactChannelArn - The Amazon Resource Name (ARN) of the contact channel.

$sel:contactArn:ContactChannel', contactChannel_contactArn - The ARN of the contact that contains the contact channel.

$sel:name:ContactChannel', contactChannel_name - The name of the contact channel.

$sel:deliveryAddress:ContactChannel', contactChannel_deliveryAddress - The details that Incident Manager uses when trying to engage the contact channel.

$sel:activationStatus:ContactChannel', contactChannel_activationStatus - A Boolean value describing if the contact channel has been activated or not. If the contact channel isn't activated, Incident Manager can't engage the contact through it.

contactChannel_type :: Lens' ContactChannel (Maybe ChannelType)

The type of the contact channel. Incident Manager supports three contact methods:

SMS

VOICE

EMAIL

contactChannel_contactChannelArn :: Lens' ContactChannel Text

The Amazon Resource Name (ARN) of the contact channel.

contactChannel_contactArn :: Lens' ContactChannel Text

The ARN of the contact that contains the contact channel.

contactChannel_name :: Lens' ContactChannel Text

The name of the contact channel.

contactChannel_deliveryAddress :: Lens' ContactChannel ContactChannelAddress

The details that Incident Manager uses when trying to engage the contact channel.

contactChannel_activationStatus :: Lens' ContactChannel ActivationStatus

A Boolean value describing if the contact channel has been activated or not. If the contact channel isn't activated, Incident Manager can't engage the contact through it.

ContactChannelAddress

data ContactChannelAddress

The details that Incident Manager uses when trying to engage the contact channel.

See: newContactChannelAddress smart constructor.

Constructors

ContactChannelAddress'

Fields

simpleAddress :: Maybe Text

The format is dependent on the type of the contact channel. The following are the expected formats:

SMS - '+' followed by the country code and phone number

VOICE - '+' followed by the country code and phone number

EMAIL - any standard email format

newContactChannelAddress :: ContactChannelAddress

Create a value of ContactChannelAddress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:simpleAddress:ContactChannelAddress', contactChannelAddress_simpleAddress - The format is dependent on the type of the contact channel. The following are the expected formats:

SMS - '+' followed by the country code and phone number

VOICE - '+' followed by the country code and phone number

EMAIL - any standard email format

contactChannelAddress_simpleAddress :: Lens' ContactChannelAddress (Maybe Text)

The format is dependent on the type of the contact channel. The following are the expected formats:

SMS - '+' followed by the country code and phone number

VOICE - '+' followed by the country code and phone number

EMAIL - any standard email format

ContactTargetInfo

data ContactTargetInfo

The contact that Incident Manager is engaging during an incident.

See: newContactTargetInfo smart constructor.

Constructors

ContactTargetInfo'

Fields

contactId :: Maybe Text

The Amazon Resource Name (ARN) of the contact.

isEssential :: Bool

A Boolean value determining if the contact's acknowledgement stops the progress of stages in the plan.

newContactTargetInfo

Arguments

:: Bool
$sel:isEssential:ContactTargetInfo'

-> ContactTargetInfo

Create a value of ContactTargetInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contactId:ContactTargetInfo', contactTargetInfo_contactId - The Amazon Resource Name (ARN) of the contact.

$sel:isEssential:ContactTargetInfo', contactTargetInfo_isEssential - A Boolean value determining if the contact's acknowledgement stops the progress of stages in the plan.

contactTargetInfo_contactId :: Lens' ContactTargetInfo (Maybe Text)

The Amazon Resource Name (ARN) of the contact.

contactTargetInfo_isEssential :: Lens' ContactTargetInfo Bool

A Boolean value determining if the contact's acknowledgement stops the progress of stages in the plan.

Engagement

data Engagement

Incident Manager reaching out to a contact or escalation plan to engage contact during an incident.

See: newEngagement smart constructor.

Constructors

Engagement'

Fields

startTime :: Maybe POSIX

The time that the engagement began.

stopTime :: Maybe POSIX

The time that the engagement ended.

incidentId :: Maybe Text

The ARN of the incident that's engaging the contact.

engagementArn :: Text

The Amazon Resource Name (ARN) of the engagement.

contactArn :: Text

The ARN of the escalation plan or contact that Incident Manager is engaging.

sender :: Text

The user that started the engagement.

newEngagement

Arguments

:: Text
$sel:engagementArn:Engagement'

-> Text
$sel:contactArn:Engagement'

-> Text
$sel:sender:Engagement'

-> Engagement

Create a value of Engagement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTime:Engagement', engagement_startTime - The time that the engagement began.

$sel:stopTime:Engagement', engagement_stopTime - The time that the engagement ended.

$sel:incidentId:Engagement', engagement_incidentId - The ARN of the incident that's engaging the contact.

$sel:engagementArn:Engagement', engagement_engagementArn - The Amazon Resource Name (ARN) of the engagement.

$sel:contactArn:Engagement', engagement_contactArn - The ARN of the escalation plan or contact that Incident Manager is engaging.

$sel:sender:Engagement', engagement_sender - The user that started the engagement.

engagement_startTime :: Lens' Engagement (Maybe UTCTime)

The time that the engagement began.

engagement_stopTime :: Lens' Engagement (Maybe UTCTime)

The time that the engagement ended.

engagement_incidentId :: Lens' Engagement (Maybe Text)

The ARN of the incident that's engaging the contact.

engagement_engagementArn :: Lens' Engagement Text

The Amazon Resource Name (ARN) of the engagement.

engagement_contactArn :: Lens' Engagement Text

The ARN of the escalation plan or contact that Incident Manager is engaging.

engagement_sender :: Lens' Engagement Text

The user that started the engagement.

Page

data Page

Incident Manager engaging a contact's contact channel.

See: newPage smart constructor.

Constructors

Page'

Fields

readTime :: Maybe POSIX

The time that the contact channel acknowledged engagement.

deliveryTime :: Maybe POSIX

The time the message was delivered to the contact channel.

incidentId :: Maybe Text

The ARN of the incident that's engaging the contact channel.

sentTime :: Maybe POSIX

The time that Incident Manager engaged the contact channel.

pageArn :: Text

The Amazon Resource Name (ARN) of the page to the contact channel.

engagementArn :: Text

The ARN of the engagement that this page is part of.

contactArn :: Text

The ARN of the contact that Incident Manager is engaging.

sender :: Text

The user that started the engagement.

newPage

Arguments

:: Text
$sel:pageArn:Page'

-> Text
$sel:engagementArn:Page'

-> Text
$sel:contactArn:Page'

-> Text
$sel:sender:Page'

-> Page

Create a value of Page with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:readTime:Page', page_readTime - The time that the contact channel acknowledged engagement.

$sel:deliveryTime:Page', page_deliveryTime - The time the message was delivered to the contact channel.

$sel:incidentId:Page', page_incidentId - The ARN of the incident that's engaging the contact channel.

$sel:sentTime:Page', page_sentTime - The time that Incident Manager engaged the contact channel.

$sel:pageArn:Page', page_pageArn - The Amazon Resource Name (ARN) of the page to the contact channel.

$sel:engagementArn:Page', page_engagementArn - The ARN of the engagement that this page is part of.

$sel:contactArn:Page', page_contactArn - The ARN of the contact that Incident Manager is engaging.

$sel:sender:Page', page_sender - The user that started the engagement.

page_readTime :: Lens' Page (Maybe UTCTime)

The time that the contact channel acknowledged engagement.

page_deliveryTime :: Lens' Page (Maybe UTCTime)

The time the message was delivered to the contact channel.

page_incidentId :: Lens' Page (Maybe Text)

The ARN of the incident that's engaging the contact channel.

page_sentTime :: Lens' Page (Maybe UTCTime)

The time that Incident Manager engaged the contact channel.

page_pageArn :: Lens' Page Text

The Amazon Resource Name (ARN) of the page to the contact channel.

page_engagementArn :: Lens' Page Text

The ARN of the engagement that this page is part of.

page_contactArn :: Lens' Page Text

The ARN of the contact that Incident Manager is engaging.

page_sender :: Lens' Page Text

The user that started the engagement.

Plan

data Plan

The stages that an escalation plan or engagement plan engages contacts and contact methods in.

See: newPlan smart constructor.

Constructors

Plan'

Fields

stages :: [Stage]

A list of stages that the escalation plan or engagement plan uses to engage contacts and contact methods.

newPlan :: Plan

Create a value of Plan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stages:Plan', plan_stages - A list of stages that the escalation plan or engagement plan uses to engage contacts and contact methods.

plan_stages :: Lens' Plan [Stage]

A list of stages that the escalation plan or engagement plan uses to engage contacts and contact methods.

Receipt

data Receipt

Records events during an engagement.

See: newReceipt smart constructor.

Constructors

Receipt'

Fields

receiptInfo :: Maybe Text

Information provided during the page acknowledgement.

contactChannelArn :: Maybe Text

The Amazon Resource Name (ARN) of the contact channel Incident Manager engaged.

receiptType :: ReceiptType

The type follows the engagement cycle, SENT, DELIVERED, and READ.

receiptTime :: POSIX

The time receipt was SENT, DELIVERED, or READ.

newReceipt

Arguments

:: ReceiptType
$sel:receiptType:Receipt'

-> UTCTime
$sel:receiptTime:Receipt'

-> Receipt

Create a value of Receipt with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:receiptInfo:Receipt', receipt_receiptInfo - Information provided during the page acknowledgement.

$sel:contactChannelArn:Receipt', receipt_contactChannelArn - The Amazon Resource Name (ARN) of the contact channel Incident Manager engaged.

$sel:receiptType:Receipt', receipt_receiptType - The type follows the engagement cycle, SENT, DELIVERED, and READ.

$sel:receiptTime:Receipt', receipt_receiptTime - The time receipt was SENT, DELIVERED, or READ.

receipt_receiptInfo :: Lens' Receipt (Maybe Text)

Information provided during the page acknowledgement.

receipt_contactChannelArn :: Lens' Receipt (Maybe Text)

The Amazon Resource Name (ARN) of the contact channel Incident Manager engaged.

receipt_receiptType :: Lens' Receipt ReceiptType

The type follows the engagement cycle, SENT, DELIVERED, and READ.

receipt_receiptTime :: Lens' Receipt UTCTime

The time receipt was SENT, DELIVERED, or READ.

Stage

data Stage

A set amount of time that an escalation plan or engagement plan engages the specified contacts or contact methods.

See: newStage smart constructor.

Constructors

Stage'

Fields

durationInMinutes :: Natural

The time to wait until beginning the next stage. The duration can only be set to 0 if a target is specified.

targets :: [Target]

The contacts or contact methods that the escalation plan or engagement plan is engaging.

newStage

Arguments

:: Natural
$sel:durationInMinutes:Stage'

-> Stage

Create a value of Stage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:durationInMinutes:Stage', stage_durationInMinutes - The time to wait until beginning the next stage. The duration can only be set to 0 if a target is specified.

$sel:targets:Stage', stage_targets - The contacts or contact methods that the escalation plan or engagement plan is engaging.

stage_durationInMinutes :: Lens' Stage Natural

The time to wait until beginning the next stage. The duration can only be set to 0 if a target is specified.

stage_targets :: Lens' Stage [Target]

The contacts or contact methods that the escalation plan or engagement plan is engaging.

Tag

data Tag

A container of a key-value name pair.

See: newTag smart constructor.

Constructors

Tag'

Fields

value :: Maybe Text

Value of the tag.

key :: Maybe Text

Name of the object key.

newTag :: Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:Tag', tag_value - Value of the tag.

$sel:key:Tag', tag_key - Name of the object key.

tag_value :: Lens' Tag (Maybe Text)

Value of the tag.

tag_key :: Lens' Tag (Maybe Text)

Name of the object key.

Target

data Target

The contact or contact channel that's being engaged.

See: newTarget smart constructor.

Constructors

Target'

Fields

channelTargetInfo :: Maybe ChannelTargetInfo

Information about the contact channel Incident Manager is engaging.

contactTargetInfo :: Maybe ContactTargetInfo

Information about the contact that Incident Manager is engaging.

newTarget :: Target

Create a value of Target with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelTargetInfo:Target', target_channelTargetInfo - Information about the contact channel Incident Manager is engaging.

$sel:contactTargetInfo:Target', target_contactTargetInfo - Information about the contact that Incident Manager is engaging.

target_channelTargetInfo :: Lens' Target (Maybe ChannelTargetInfo)

Information about the contact channel Incident Manager is engaging.

target_contactTargetInfo :: Lens' Target (Maybe ContactTargetInfo)

Information about the contact that Incident Manager is engaging.

TimeRange

data TimeRange

A range of between two set times

See: newTimeRange smart constructor.

Constructors

TimeRange'

Fields

startTime :: Maybe POSIX

The start of the time range.

endTime :: Maybe POSIX

The end of the time range.

newTimeRange :: TimeRange

Create a value of TimeRange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTime:TimeRange', timeRange_startTime - The start of the time range.

$sel:endTime:TimeRange', timeRange_endTime - The end of the time range.

timeRange_startTime :: Lens' TimeRange (Maybe UTCTime)

The start of the time range.

timeRange_endTime :: Lens' TimeRange (Maybe UTCTime)

The end of the time range.