Network.AWS.SSMContacts.Types.Engagement

Description

Documentation

data Engagement

Incident Manager reaching out to a contact or escalation plan to engage contact during an incident.

See: newEngagement smart constructor.

Constructors

Engagement'

Fields

startTime :: Maybe POSIX

The time that the engagement began.

stopTime :: Maybe POSIX

The time that the engagement ended.

incidentId :: Maybe Text

The ARN of the incident that's engaging the contact.

engagementArn :: Text

The Amazon Resource Name (ARN) of the engagement.

contactArn :: Text

The ARN of the escalation plan or contact that Incident Manager is engaging.

sender :: Text

The user that started the engagement.

newEngagement

Arguments

:: Text
$sel:engagementArn:Engagement'

-> Text
$sel:contactArn:Engagement'

-> Text
$sel:sender:Engagement'

-> Engagement

Create a value of Engagement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTime:Engagement', engagement_startTime - The time that the engagement began.

$sel:stopTime:Engagement', engagement_stopTime - The time that the engagement ended.

$sel:incidentId:Engagement', engagement_incidentId - The ARN of the incident that's engaging the contact.

$sel:engagementArn:Engagement', engagement_engagementArn - The Amazon Resource Name (ARN) of the engagement.

$sel:contactArn:Engagement', engagement_contactArn - The ARN of the escalation plan or contact that Incident Manager is engaging.

$sel:sender:Engagement', engagement_sender - The user that started the engagement.

engagement_startTime :: Lens' Engagement (Maybe UTCTime)

The time that the engagement began.

engagement_stopTime :: Lens' Engagement (Maybe UTCTime)

The time that the engagement ended.

engagement_incidentId :: Lens' Engagement (Maybe Text)

The ARN of the incident that's engaging the contact.

engagement_engagementArn :: Lens' Engagement Text

The Amazon Resource Name (ARN) of the engagement.

engagement_contactArn :: Lens' Engagement Text

The ARN of the escalation plan or contact that Incident Manager is engaging.

engagement_sender :: Lens' Engagement Text

The user that started the engagement.