Network.AWS.SSMContacts.Types.Page

Description

Documentation

data Page

Incident Manager engaging a contact's contact channel.

See: newPage smart constructor.

Constructors

Page'

Fields

readTime :: Maybe POSIX

The time that the contact channel acknowledged engagement.

deliveryTime :: Maybe POSIX

The time the message was delivered to the contact channel.

incidentId :: Maybe Text

The ARN of the incident that's engaging the contact channel.

sentTime :: Maybe POSIX

The time that Incident Manager engaged the contact channel.

pageArn :: Text

The Amazon Resource Name (ARN) of the page to the contact channel.

engagementArn :: Text

The ARN of the engagement that this page is part of.

contactArn :: Text

The ARN of the contact that Incident Manager is engaging.

sender :: Text

The user that started the engagement.

newPage

Arguments

:: Text
$sel:pageArn:Page'

-> Text
$sel:engagementArn:Page'

-> Text
$sel:contactArn:Page'

-> Text
$sel:sender:Page'

-> Page

Create a value of Page with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:readTime:Page', page_readTime - The time that the contact channel acknowledged engagement.

$sel:deliveryTime:Page', page_deliveryTime - The time the message was delivered to the contact channel.

$sel:incidentId:Page', page_incidentId - The ARN of the incident that's engaging the contact channel.

$sel:sentTime:Page', page_sentTime - The time that Incident Manager engaged the contact channel.

$sel:pageArn:Page', page_pageArn - The Amazon Resource Name (ARN) of the page to the contact channel.

$sel:engagementArn:Page', page_engagementArn - The ARN of the engagement that this page is part of.

$sel:contactArn:Page', page_contactArn - The ARN of the contact that Incident Manager is engaging.

$sel:sender:Page', page_sender - The user that started the engagement.

page_readTime :: Lens' Page (Maybe UTCTime)

The time that the contact channel acknowledged engagement.

page_deliveryTime :: Lens' Page (Maybe UTCTime)

The time the message was delivered to the contact channel.

page_incidentId :: Lens' Page (Maybe Text)

The ARN of the incident that's engaging the contact channel.

page_sentTime :: Lens' Page (Maybe UTCTime)

The time that Incident Manager engaged the contact channel.

page_pageArn :: Lens' Page Text

The Amazon Resource Name (ARN) of the page to the contact channel.

page_engagementArn :: Lens' Page Text

The ARN of the engagement that this page is part of.

page_contactArn :: Lens' Page Text

The ARN of the contact that Incident Manager is engaging.

page_sender :: Lens' Page Text

The user that started the engagement.