Network.AWS.SSMContacts.Types.Receipt

Description

Documentation

data Receipt

Records events during an engagement.

See: newReceipt smart constructor.

Constructors

Receipt'

Fields

receiptInfo :: Maybe Text

Information provided during the page acknowledgement.

contactChannelArn :: Maybe Text

The Amazon Resource Name (ARN) of the contact channel Incident Manager engaged.

receiptType :: ReceiptType

The type follows the engagement cycle, SENT, DELIVERED, and READ.

receiptTime :: POSIX

The time receipt was SENT, DELIVERED, or READ.

newReceipt

Arguments

:: ReceiptType
$sel:receiptType:Receipt'

-> UTCTime
$sel:receiptTime:Receipt'

-> Receipt

Create a value of Receipt with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:receiptInfo:Receipt', receipt_receiptInfo - Information provided during the page acknowledgement.

$sel:contactChannelArn:Receipt', receipt_contactChannelArn - The Amazon Resource Name (ARN) of the contact channel Incident Manager engaged.

$sel:receiptType:Receipt', receipt_receiptType - The type follows the engagement cycle, SENT, DELIVERED, and READ.

$sel:receiptTime:Receipt', receipt_receiptTime - The time receipt was SENT, DELIVERED, or READ.

receipt_receiptInfo :: Lens' Receipt (Maybe Text)

Information provided during the page acknowledgement.

receipt_contactChannelArn :: Lens' Receipt (Maybe Text)

The Amazon Resource Name (ARN) of the contact channel Incident Manager engaged.

receipt_receiptType :: Lens' Receipt ReceiptType

The type follows the engagement cycle, SENT, DELIVERED, and READ.

receipt_receiptTime :: Lens' Receipt UTCTime

The time receipt was SENT, DELIVERED, or READ.