Network.AWS.SSMContacts.UpdateContactChannel

Description

Updates a contact's contact channel.

Creating a Request

data UpdateContactChannel

See: newUpdateContactChannel smart constructor.

Constructors

UpdateContactChannel'

Fields

name :: Maybe Text

The name of the contact channel.

deliveryAddress :: Maybe ContactChannelAddress

The details that Incident Manager uses when trying to engage the contact channel.

contactChannelId :: Text

The Amazon Resource Name (ARN) of the contact channel you want to update.

newUpdateContactChannel

Arguments

:: Text
$sel:contactChannelId:UpdateContactChannel'

-> UpdateContactChannel

Create a value of UpdateContactChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateContactChannel', updateContactChannel_name - The name of the contact channel.

$sel:deliveryAddress:UpdateContactChannel', updateContactChannel_deliveryAddress - The details that Incident Manager uses when trying to engage the contact channel.

$sel:contactChannelId:UpdateContactChannel', updateContactChannel_contactChannelId - The Amazon Resource Name (ARN) of the contact channel you want to update.

Request Lenses

updateContactChannel_name :: Lens' UpdateContactChannel (Maybe Text)

The name of the contact channel.

updateContactChannel_deliveryAddress :: Lens' UpdateContactChannel (Maybe ContactChannelAddress)

The details that Incident Manager uses when trying to engage the contact channel.

updateContactChannel_contactChannelId :: Lens' UpdateContactChannel Text

The Amazon Resource Name (ARN) of the contact channel you want to update.

Destructuring the Response

data UpdateContactChannelResponse

See: newUpdateContactChannelResponse smart constructor.

Constructors

UpdateContactChannelResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateContactChannelResponse

Arguments

:: Int
$sel:httpStatus:UpdateContactChannelResponse'

-> UpdateContactChannelResponse

Create a value of UpdateContactChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateContactChannelResponse', updateContactChannelResponse_httpStatus - The response's http status code.

Response Lenses

updateContactChannelResponse_httpStatus :: Lens' UpdateContactChannelResponse Int

The response's http status code.