Network.AWS.SSMIncidents.Lens

Description

Operations

DeleteReplicationSet

deleteReplicationSet_arn :: Lens' DeleteReplicationSet Text

The Amazon Resource Name (ARN) of the replication set you're deleting.

deleteReplicationSetResponse_httpStatus :: Lens' DeleteReplicationSetResponse Int

The response's http status code.

UpdateReplicationSet

updateReplicationSet_clientToken :: Lens' UpdateReplicationSet (Maybe Text)

A token ensuring that the action is called only once with the specified details.

updateReplicationSet_actions :: Lens' UpdateReplicationSet (NonEmpty UpdateReplicationSetAction)

An action to add or delete a Region.

updateReplicationSet_arn :: Lens' UpdateReplicationSet Text

The Amazon Resource Name (ARN) of the replication set you're updating.

updateReplicationSetResponse_httpStatus :: Lens' UpdateReplicationSetResponse Int

The response's http status code.

ListReplicationSets

listReplicationSets_nextToken :: Lens' ListReplicationSets (Maybe Text)

The pagination token to continue to the next page of results.

listReplicationSets_maxResults :: Lens' ListReplicationSets (Maybe Natural)

The maximum number of results per page.

listReplicationSetsResponse_nextToken :: Lens' ListReplicationSetsResponse (Maybe Text)

The pagination token to continue to the next page of results.

listReplicationSetsResponse_httpStatus :: Lens' ListReplicationSetsResponse Int

The response's http status code.

listReplicationSetsResponse_replicationSetArns :: Lens' ListReplicationSetsResponse [Text]

The Amazon Resource Name (ARN) of the list replication set.

UpdateIncidentRecord

updateIncidentRecord_summary :: Lens' UpdateIncidentRecord (Maybe Text)

The summary describes what has happened during the incident.

updateIncidentRecord_status :: Lens' UpdateIncidentRecord (Maybe IncidentRecordStatus)

The status of the incident. An incident can be Open or Resolved.

updateIncidentRecord_notificationTargets :: Lens' UpdateIncidentRecord (Maybe [NotificationTargetItem])

The SNS targets that are notified when updates are made to an incident.

Using multiple SNS topics creates redundancy in the case that a Region is down during the incident.

updateIncidentRecord_clientToken :: Lens' UpdateIncidentRecord (Maybe Text)

A token ensuring that the action is called only once with the specified details.

updateIncidentRecord_impact :: Lens' UpdateIncidentRecord (Maybe Natural)

Defines the impact to customers and applications. Providing an impact overwrites the impact provided by the response plan.

Possible impacts:

1 - Critical impact, this typically relates to full application failure that impacts many to all customers.

2 - High impact, partial application failure with impact to many customers.

3 - Medium impact, the application is providing reduced service to customers.

4 - Low impact, customer might aren't impacted by the problem yet.

5 - No impact, customers aren't currently impacted but urgent action is needed to avoid impact.

updateIncidentRecord_chatChannel :: Lens' UpdateIncidentRecord (Maybe ChatChannel)

The AWS Chatbot chat channel for responders to collaborate in.

updateIncidentRecord_title :: Lens' UpdateIncidentRecord (Maybe Text)

The title of the incident is a brief and easily recognizable.

updateIncidentRecord_arn :: Lens' UpdateIncidentRecord Text

The Amazon Resource Name (ARN) of the incident record you are updating.

updateIncidentRecordResponse_httpStatus :: Lens' UpdateIncidentRecordResponse Int

The response's http status code.

DeleteIncidentRecord

deleteIncidentRecord_arn :: Lens' DeleteIncidentRecord Text

The Amazon Resource Name (ARN) of the incident record you are deleting.

deleteIncidentRecordResponse_httpStatus :: Lens' DeleteIncidentRecordResponse Int

The response's http status code.

CreateReplicationSet

createReplicationSet_clientToken :: Lens' CreateReplicationSet (Maybe Text)

A token ensuring that the action is called only once with the specified details.

createReplicationSet_regions :: Lens' CreateReplicationSet (HashMap Text RegionMapInputValue)

The Regions that Incident Manager replicates your data to. You can have up to three Regions in your replication set.

createReplicationSetResponse_httpStatus :: Lens' CreateReplicationSetResponse Int

The response's http status code.

createReplicationSetResponse_arn :: Lens' CreateReplicationSetResponse Text

The Amazon Resource Name (ARN) of the replication set.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the response plan.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (HashMap Text Text)

A list of tags for the response plan.

GetResourcePolicies

getResourcePolicies_nextToken :: Lens' GetResourcePolicies (Maybe Text)

The pagination token to continue to the next page of results.

getResourcePolicies_maxResults :: Lens' GetResourcePolicies (Maybe Natural)

The maximum number of resource policies to display per page of results.

getResourcePolicies_resourceArn :: Lens' GetResourcePolicies Text

The Amazon Resource Name (ARN) of the response plan with the attached resource policy.

getResourcePoliciesResponse_nextToken :: Lens' GetResourcePoliciesResponse (Maybe Text)

The pagination token to continue to the next page of results.

getResourcePoliciesResponse_httpStatus :: Lens' GetResourcePoliciesResponse Int

The response's http status code.

getResourcePoliciesResponse_resourcePolicies :: Lens' GetResourcePoliciesResponse [ResourcePolicy]

Details about the resource policy attached to the response plan.

GetIncidentRecord

getIncidentRecord_arn :: Lens' GetIncidentRecord Text

The Amazon Resource Name (ARN) of the incident record.

getIncidentRecordResponse_httpStatus :: Lens' GetIncidentRecordResponse Int

The response's http status code.

getIncidentRecordResponse_incidentRecord :: Lens' GetIncidentRecordResponse IncidentRecord

Details structure of the incident record.

GetReplicationSet

getReplicationSet_arn :: Lens' GetReplicationSet Text

The Amazon Resource Name (ARN) of the replication set you want to retrieve.

getReplicationSetResponse_httpStatus :: Lens' GetReplicationSetResponse Int

The response's http status code.

getReplicationSetResponse_replicationSet :: Lens' GetReplicationSetResponse ReplicationSet

Details of the replication set.

ListRelatedItems

listRelatedItems_nextToken :: Lens' ListRelatedItems (Maybe Text)

The pagination token to continue to the next page of results.

listRelatedItems_maxResults :: Lens' ListRelatedItems (Maybe Natural)

The maximum number of related items per page.

listRelatedItems_incidentRecordArn :: Lens' ListRelatedItems Text

The Amazon Resource Name (ARN) of the incident record that you are listing related items for.

listRelatedItemsResponse_nextToken :: Lens' ListRelatedItemsResponse (Maybe Text)

The pagination token to continue to the next page of results.

listRelatedItemsResponse_httpStatus :: Lens' ListRelatedItemsResponse Int

The response's http status code.

listRelatedItemsResponse_relatedItems :: Lens' ListRelatedItemsResponse [RelatedItem]

Details about each related item.

UpdateDeletionProtection

updateDeletionProtection_clientToken :: Lens' UpdateDeletionProtection (Maybe Text)

A token ensuring that the action is called only once with the specified details.

updateDeletionProtection_arn :: Lens' UpdateDeletionProtection Text

The Amazon Resource Name (ARN) of the replication set you're updating.

updateDeletionProtection_deletionProtected :: Lens' UpdateDeletionProtection Bool

Details if deletion protection is enabled or disabled in your account.

updateDeletionProtectionResponse_httpStatus :: Lens' UpdateDeletionProtectionResponse Int

The response's http status code.

GetResponsePlan

getResponsePlan_arn :: Lens' GetResponsePlan Text

The Amazon Resource Name (ARN) of the response plan.

getResponsePlanResponse_actions :: Lens' GetResponsePlanResponse (Maybe [Action])

The actions that this response plan takes at the beginning of the incident.

getResponsePlanResponse_displayName :: Lens' GetResponsePlanResponse (Maybe Text)

The long format name of the response plan. Can contain spaces.

getResponsePlanResponse_chatChannel :: Lens' GetResponsePlanResponse (Maybe ChatChannel)

The AWS Chatbot chat channel used for collaboration during an incident.

getResponsePlanResponse_engagements :: Lens' GetResponsePlanResponse (Maybe [Text])

The contacts and escalation plans that the response plan engages during an incident.

getResponsePlanResponse_httpStatus :: Lens' GetResponsePlanResponse Int

The response's http status code.

getResponsePlanResponse_arn :: Lens' GetResponsePlanResponse Text

The ARN of the response plan.

getResponsePlanResponse_incidentTemplate :: Lens' GetResponsePlanResponse IncidentTemplate

Details used to create the incident when using this response plan.

getResponsePlanResponse_name :: Lens' GetResponsePlanResponse Text

The short format name of the response plan. Can't contain spaces.

CreateResponsePlan

createResponsePlan_clientToken :: Lens' CreateResponsePlan (Maybe Text)

A token ensuring that the action is called only once with the specified details.

createResponsePlan_actions :: Lens' CreateResponsePlan (Maybe [Action])

The actions that the response plan starts at the beginning of an incident.

createResponsePlan_displayName :: Lens' CreateResponsePlan (Maybe Text)

The long format of the response plan name. This field can contain spaces.

createResponsePlan_chatChannel :: Lens' CreateResponsePlan (Maybe ChatChannel)

The AWS Chatbot chat channel used for collaboration during an incident.

createResponsePlan_engagements :: Lens' CreateResponsePlan (Maybe [Text])

The contacts and escalation plans that the response plan engages during an incident.

createResponsePlan_tags :: Lens' CreateResponsePlan (Maybe (HashMap Text Text))

A list of tags that you are adding to the response plan.

createResponsePlan_incidentTemplate :: Lens' CreateResponsePlan IncidentTemplate

Details used to create an incident when using this response plan.

createResponsePlan_name :: Lens' CreateResponsePlan Text

The short format name of the response plan. Can't include spaces.

createResponsePlanResponse_httpStatus :: Lens' CreateResponsePlanResponse Int

The response's http status code.

createResponsePlanResponse_arn :: Lens' CreateResponsePlanResponse Text

The Amazon Resource Name (ARN) of the response plan.

ListIncidentRecords

listIncidentRecords_filters :: Lens' ListIncidentRecords (Maybe [Filter])

Filter the list of incident records you are searching through. You can filter on the following keys:

creationTime

impact

status

createdBy

listIncidentRecords_nextToken :: Lens' ListIncidentRecords (Maybe Text)

The pagination token to continue to the next page of results.

listIncidentRecords_maxResults :: Lens' ListIncidentRecords (Maybe Natural)

The maximum number of results per page.

listIncidentRecordsResponse_nextToken :: Lens' ListIncidentRecordsResponse (Maybe Text)

The pagination token to continue to the next page of results.

listIncidentRecordsResponse_httpStatus :: Lens' ListIncidentRecordsResponse Int

The response's http status code.

listIncidentRecordsResponse_incidentRecordSummaries :: Lens' ListIncidentRecordsResponse [IncidentRecordSummary]

The details of each listed incident record.

UpdateRelatedItems

updateRelatedItems_clientToken :: Lens' UpdateRelatedItems (Maybe Text)

A token ensuring that the action is called only once with the specified details.

updateRelatedItems_incidentRecordArn :: Lens' UpdateRelatedItems Text

The Amazon Resource Name (ARN) of the incident record you are updating related items in.

updateRelatedItems_relatedItemsUpdate :: Lens' UpdateRelatedItems RelatedItemsUpdate

Details about the item you are adding or deleting.

updateRelatedItemsResponse_httpStatus :: Lens' UpdateRelatedItemsResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the response plan you're adding the tags to.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

A list of tags that you are adding to the response plan.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

PutResourcePolicy

putResourcePolicy_policy :: Lens' PutResourcePolicy Text

Details of the resource policy.

putResourcePolicy_resourceArn :: Lens' PutResourcePolicy Text

The Amazon Resource Name (ARN) of the response plan you're adding the resource policy to.

putResourcePolicyResponse_httpStatus :: Lens' PutResourcePolicyResponse Int

The response's http status code.

putResourcePolicyResponse_policyId :: Lens' PutResourcePolicyResponse Text

The ID of the resource policy.

DeleteResourcePolicy

deleteResourcePolicy_policyId :: Lens' DeleteResourcePolicy Text

The ID of the resource policy you're deleting.

deleteResourcePolicy_resourceArn :: Lens' DeleteResourcePolicy Text

The Amazon Resource Name (ARN) of the resource you're deleting the policy from.

deleteResourcePolicyResponse_httpStatus :: Lens' DeleteResourcePolicyResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the response plan you're removing a tag from.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

The name of the tag you're removing from the response plan.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

CreateTimelineEvent

createTimelineEvent_clientToken :: Lens' CreateTimelineEvent (Maybe Text)

A token ensuring that the action is called only once with the specified details.

createTimelineEvent_eventData :: Lens' CreateTimelineEvent Text

A valid JSON string. There is no other schema imposed. A short description of the event.

createTimelineEvent_eventTime :: Lens' CreateTimelineEvent UTCTime

The time that the event occurred.

createTimelineEvent_eventType :: Lens' CreateTimelineEvent Text

The type of the event. You can create timeline events of type Custom Event.

createTimelineEvent_incidentRecordArn :: Lens' CreateTimelineEvent Text

The Amazon Resource Name (ARN) of the incident record you are adding the event to.

createTimelineEventResponse_httpStatus :: Lens' CreateTimelineEventResponse Int

The response's http status code.

createTimelineEventResponse_eventId :: Lens' CreateTimelineEventResponse Text

The ID of the event for easy reference later.

createTimelineEventResponse_incidentRecordArn :: Lens' CreateTimelineEventResponse Text

The ARN of the incident record that you added the event to.

ListTimelineEvents

listTimelineEvents_filters :: Lens' ListTimelineEvents (Maybe [Filter])

Filters the timeline events based on the provided conditional values. You can filter timeline events using the following keys:

eventTime

eventType

listTimelineEvents_sortOrder :: Lens' ListTimelineEvents (Maybe SortOrder)

Sorts the order of timeline events by the value specified in the sortBy field.

listTimelineEvents_nextToken :: Lens' ListTimelineEvents (Maybe Text)

The pagination token to continue to the next page of results.

listTimelineEvents_maxResults :: Lens' ListTimelineEvents (Maybe Natural)

The maximum number of results per page.

listTimelineEvents_sortBy :: Lens' ListTimelineEvents (Maybe TimelineEventSort)

Sort by the specified key value pair.

listTimelineEvents_incidentRecordArn :: Lens' ListTimelineEvents Text

The Amazon Resource Name (ARN) of the incident that the event is part of.

listTimelineEventsResponse_nextToken :: Lens' ListTimelineEventsResponse (Maybe Text)

The pagination token to continue to the next page of results.

listTimelineEventsResponse_httpStatus :: Lens' ListTimelineEventsResponse Int

The response's http status code.

listTimelineEventsResponse_eventSummaries :: Lens' ListTimelineEventsResponse [EventSummary]

Details about each event that occurred during the incident.

StartIncident

startIncident_clientToken :: Lens' StartIncident (Maybe Text)

A token ensuring that the action is called only once with the specified details.

startIncident_triggerDetails :: Lens' StartIncident (Maybe TriggerDetails)

Details of what created the incident record in Incident Manager.

startIncident_relatedItems :: Lens' StartIncident (Maybe [RelatedItem])

Add related items to the incident for other responders to use. Related items are AWS resources, external links, or files uploaded to an S3 bucket.

startIncident_impact :: Lens' StartIncident (Maybe Natural)

Defines the impact to the customers. Providing an impact overwrites the impact provided by a response plan.

Possible impacts:

1 - Critical impact, this typically relates to full application failure that impacts many to all customers.

2 - High impact, partial application failure with impact to many customers.

3 - Medium impact, the application is providing reduced service to customers.

4 - Low impact, customer might aren't impacted by the problem yet.

5 - No impact, customers aren't currently impacted but urgent action is needed to avoid impact.

startIncident_title :: Lens' StartIncident (Maybe Text)

Provide a title for the incident. Providing a title overwrites the title provided by the response plan.

startIncident_responsePlanArn :: Lens' StartIncident Text

The Amazon Resource Name (ARN) of the response plan that pre-defines summary, chat channels, SNS topics, runbooks, title, and impact of the incident.

startIncidentResponse_httpStatus :: Lens' StartIncidentResponse Int

The response's http status code.

startIncidentResponse_incidentRecordArn :: Lens' StartIncidentResponse Text

The ARN of the newly created incident record.

DeleteTimelineEvent

deleteTimelineEvent_eventId :: Lens' DeleteTimelineEvent Text

The ID of the event you are updating. You can find this by using ListTimelineEvents.

deleteTimelineEvent_incidentRecordArn :: Lens' DeleteTimelineEvent Text

The Amazon Resource Name (ARN) of the incident that the event is part of.

deleteTimelineEventResponse_httpStatus :: Lens' DeleteTimelineEventResponse Int

The response's http status code.

UpdateTimelineEvent

updateTimelineEvent_eventData :: Lens' UpdateTimelineEvent (Maybe Text)

A short description of the event.

updateTimelineEvent_clientToken :: Lens' UpdateTimelineEvent (Maybe Text)

A token ensuring that the action is called only once with the specified details.

updateTimelineEvent_eventTime :: Lens' UpdateTimelineEvent (Maybe UTCTime)

The time that the event occurred.

updateTimelineEvent_eventType :: Lens' UpdateTimelineEvent (Maybe Text)

The type of the event. You can update events of type Custom Event.

updateTimelineEvent_eventId :: Lens' UpdateTimelineEvent Text

The ID of the event you are updating. You can find this by using ListTimelineEvents.

updateTimelineEvent_incidentRecordArn :: Lens' UpdateTimelineEvent Text

The Amazon Resource Name (ARN) of the incident that the timeline event is part of.

updateTimelineEventResponse_httpStatus :: Lens' UpdateTimelineEventResponse Int

The response's http status code.

ListResponsePlans

listResponsePlans_nextToken :: Lens' ListResponsePlans (Maybe Text)

The pagination token to continue to the next page of results.

listResponsePlans_maxResults :: Lens' ListResponsePlans (Maybe Natural)

The maximum number of response plans per page.

listResponsePlansResponse_nextToken :: Lens' ListResponsePlansResponse (Maybe Text)

The pagination token to continue to the next page of results.

listResponsePlansResponse_httpStatus :: Lens' ListResponsePlansResponse Int

The response's http status code.

listResponsePlansResponse_responsePlanSummaries :: Lens' ListResponsePlansResponse [ResponsePlanSummary]

Details of each response plan.

GetTimelineEvent

getTimelineEvent_eventId :: Lens' GetTimelineEvent Text

The ID of the event. You can get an event's ID when you create it or by using ListTimelineEvents.

getTimelineEvent_incidentRecordArn :: Lens' GetTimelineEvent Text

The Amazon Resource Name (ARN) of the incident that the timeline event is part of.

getTimelineEventResponse_httpStatus :: Lens' GetTimelineEventResponse Int

The response's http status code.

getTimelineEventResponse_event :: Lens' GetTimelineEventResponse TimelineEvent

Details about the timeline event.

UpdateResponsePlan

updateResponsePlan_incidentTemplateImpact :: Lens' UpdateResponsePlan (Maybe Natural)

Defines the impact to the customers. Providing an impact overwrites the impact provided by a response plan.

Possible impacts:

5 - Severe impact

4 - High impact

3 - Medium impact

2 - Low impact

1 - No impact

updateResponsePlan_clientToken :: Lens' UpdateResponsePlan (Maybe Text)

A token ensuring that the action is called only once with the specified details.

updateResponsePlan_actions :: Lens' UpdateResponsePlan (Maybe [Action])

The actions that this response plan takes at the beginning of an incident.

updateResponsePlan_incidentTemplateSummary :: Lens' UpdateResponsePlan (Maybe Text)

A brief summary of the incident. This typically contains what has happened, what's currently happening, and next steps.

updateResponsePlan_displayName :: Lens' UpdateResponsePlan (Maybe Text)

The long format name of the response plan. Can't contain spaces.

updateResponsePlan_chatChannel :: Lens' UpdateResponsePlan (Maybe ChatChannel)

The AWS Chatbot chat channel used for collaboration during an incident.

Use the empty structure to remove the chat channel from the response plan.

updateResponsePlan_incidentTemplateDedupeString :: Lens' UpdateResponsePlan (Maybe Text)

Used to create only one incident record for an incident.

updateResponsePlan_incidentTemplateTitle :: Lens' UpdateResponsePlan (Maybe Text)

The short format name of the incident. Can't contain spaces.

updateResponsePlan_engagements :: Lens' UpdateResponsePlan (Maybe [Text])

The contacts and escalation plans that Incident Manager engages at the start of the incident.

updateResponsePlan_incidentTemplateNotificationTargets :: Lens' UpdateResponsePlan (Maybe [NotificationTargetItem])

The SNS targets that are notified when updates are made to an incident.

updateResponsePlan_arn :: Lens' UpdateResponsePlan Text

The Amazon Resource Name (ARN) of the response plan.

updateResponsePlanResponse_httpStatus :: Lens' UpdateResponsePlanResponse Int

The response's http status code.

DeleteResponsePlan

deleteResponsePlan_arn :: Lens' DeleteResponsePlan Text

The Amazon Resource Name (ARN) of the response plan.

deleteResponsePlanResponse_httpStatus :: Lens' DeleteResponsePlanResponse Int

The response's http status code.

Types

Action

action_ssmAutomation :: Lens' Action (Maybe SsmAutomation)

The Systems Manager automation document to start as the runbook at the beginning of the incident.

AddRegionAction

addRegionAction_sseKmsKeyId :: Lens' AddRegionAction (Maybe Text)

The KMS key ID to use to encrypt your replication set.

addRegionAction_regionName :: Lens' AddRegionAction Text

The Region name to add to the replication set.

AttributeValueList

attributeValueList_stringValues :: Lens' AttributeValueList (Maybe [Text])

The list of string values that the filter matches.

attributeValueList_integerValues :: Lens' AttributeValueList (Maybe [Int])

The list of integer values that the filter matches.

AutomationExecution

automationExecution_ssmExecutionArn :: Lens' AutomationExecution (Maybe Text)

The Amazon Resource Name (ARN) of the automation process.

ChatChannel

chatChannel_empty :: Lens' ChatChannel (Maybe EmptyChatChannel)

Used to remove the chat channel from an incident record or response plan.

chatChannel_chatbotSns :: Lens' ChatChannel (Maybe (NonEmpty Text))

The SNS targets that AWS Chatbot uses to notify the chat channel of updates to an incident. You can also make updates to the incident through the chat channel by using the SNS topics.

Condition

condition_after :: Lens' Condition (Maybe UTCTime)

After the specified timestamp.

condition_equals :: Lens' Condition (Maybe AttributeValueList)

The value is equal to the provided string or integer.

condition_before :: Lens' Condition (Maybe UTCTime)

Before the specified timestamp

DeleteRegionAction

deleteRegionAction_regionName :: Lens' DeleteRegionAction Text

The name of the Region you're deleting from the replication set.

EmptyChatChannel

EventSummary

eventSummary_eventId :: Lens' EventSummary Text

The timeline event ID.

eventSummary_eventTime :: Lens' EventSummary UTCTime

The time that the event occurred.

eventSummary_eventType :: Lens' EventSummary Text

The type of event. The timeline event must be Custom Event.

eventSummary_eventUpdatedTime :: Lens' EventSummary UTCTime

The time that the timeline event was last updated.

eventSummary_incidentRecordArn :: Lens' EventSummary Text

The Amazon Resource Name (ARN) of the incident that the event happened during.

Filter

filter_condition :: Lens' Filter Condition

The condition accepts before or after a specified time, equal to a string, or equal to an integer.

filter_key :: Lens' Filter Text

The key that you're filtering on.

IncidentRecord

incidentRecord_summary :: Lens' IncidentRecord (Maybe Text)

The summary of the incident. The summary is a brief synopsis of what occurred, what is currently happening, and context.

incidentRecord_notificationTargets :: Lens' IncidentRecord (Maybe [NotificationTargetItem])

The SNS targets that are notified when updates are made to an incident.

incidentRecord_resolvedTime :: Lens' IncidentRecord (Maybe UTCTime)

The time at which the incident was resolved. This appears as a timeline event.

incidentRecord_chatChannel :: Lens' IncidentRecord (Maybe ChatChannel)

The chat channel used for collaboration during an incident.

incidentRecord_automationExecutions :: Lens' IncidentRecord (Maybe [AutomationExecution])

The runbook, or automation document, that's run at the beginning of the incident.

incidentRecord_arn :: Lens' IncidentRecord Text

The Amazon Resource Name (ARN) of the incident record.

incidentRecord_creationTime :: Lens' IncidentRecord UTCTime

The time that Incident Manager created the incident record.

incidentRecord_dedupeString :: Lens' IncidentRecord Text

The string Incident Manager uses to prevent duplicate incidents from being created by the same incident.

incidentRecord_impact :: Lens' IncidentRecord Natural

The impact of the incident on customers and applications.

incidentRecord_incidentRecordSource :: Lens' IncidentRecord IncidentRecordSource

Details about the action that started the incident.

incidentRecord_lastModifiedBy :: Lens' IncidentRecord Text

Who modified the incident most recently.

incidentRecord_lastModifiedTime :: Lens' IncidentRecord UTCTime

The time at which the incident was most recently modified.

incidentRecord_status :: Lens' IncidentRecord IncidentRecordStatus

The current status of the incident.

incidentRecord_title :: Lens' IncidentRecord Text

The title of the incident.

IncidentRecordSource

incidentRecordSource_invokedBy :: Lens' IncidentRecordSource (Maybe Text)

The principal the assumed the role specified of the createdBy.

incidentRecordSource_resourceArn :: Lens' IncidentRecordSource (Maybe Text)

The resource that caused the incident to be created.

incidentRecordSource_createdBy :: Lens' IncidentRecordSource Text

The principal that started the incident.

incidentRecordSource_source :: Lens' IncidentRecordSource Text

The service that started the incident. This can be manually created from Incident Manager, automatically created using an AWS CloudWatch alarm, or Amazon EventBridge event.

IncidentRecordSummary

incidentRecordSummary_resolvedTime :: Lens' IncidentRecordSummary (Maybe UTCTime)

The time the incident was resolved.

incidentRecordSummary_arn :: Lens' IncidentRecordSummary Text

The Amazon Resource Name (ARN) of the incident.

incidentRecordSummary_creationTime :: Lens' IncidentRecordSummary UTCTime

The time the incident was created.

incidentRecordSummary_impact :: Lens' IncidentRecordSummary Natural

Defines the impact to customers and applications.

incidentRecordSummary_incidentRecordSource :: Lens' IncidentRecordSummary IncidentRecordSource

What caused Incident Manager to create the incident.

incidentRecordSummary_status :: Lens' IncidentRecordSummary IncidentRecordStatus

The current status of the incident.

incidentRecordSummary_title :: Lens' IncidentRecordSummary Text

The title of the incident. This value is either provided by the response plan or overwritten on creation.

IncidentTemplate

incidentTemplate_summary :: Lens' IncidentTemplate (Maybe Text)

The summary of the incident. The summary is a brief synopsis of what occurred, what's currently happening, and context.

incidentTemplate_notificationTargets :: Lens' IncidentTemplate (Maybe [NotificationTargetItem])

The SNS targets that are notified when updates are made to an incident.

incidentTemplate_dedupeString :: Lens' IncidentTemplate (Maybe Text)

Used to stop Incident Manager from creating multiple incident records for the same incident.

incidentTemplate_impact :: Lens' IncidentTemplate Natural

The impact of the incident on your customers and applications.

incidentTemplate_title :: Lens' IncidentTemplate Text

The title of the incident.

ItemIdentifier

itemIdentifier_type :: Lens' ItemIdentifier ItemType

The type of related item. Incident Manager supports the following types:

ANALYSIS

INCIDENT

METRIC

PARENT

ATTACHMENT

OTHER

itemIdentifier_value :: Lens' ItemIdentifier ItemValue

Details about the related item.

ItemValue

itemValue_arn :: Lens' ItemValue (Maybe Text)

The Amazon Resource Name (ARN) of the related item, if the related item is an Amazon resource.

itemValue_url :: Lens' ItemValue (Maybe Text)

The URL, if the related item is a non-AWS resource.

itemValue_metricDefinition :: Lens' ItemValue (Maybe Text)

The metric definition, if the related item is a metric in CloudWatch.

NotificationTargetItem

notificationTargetItem_snsTopicArn :: Lens' NotificationTargetItem (Maybe Text)

The Amazon Resource Name (ARN) of the SNS topic.

RegionInfo

regionInfo_statusMessage :: Lens' RegionInfo (Maybe Text)

Information displayed about the status of the Region.

regionInfo_sseKmsKeyId :: Lens' RegionInfo (Maybe Text)

The ID of the KMS key used to encrypt the data in this Region.

regionInfo_status :: Lens' RegionInfo RegionStatus

The status of the Region in the replication set.

regionInfo_statusUpdateDateTime :: Lens' RegionInfo UTCTime

The most recent date and time that the Region's status was updated.

RegionMapInputValue

regionMapInputValue_sseKmsKeyId :: Lens' RegionMapInputValue (Maybe Text)

The KMS key used to encrypt the data in your replication set.

RelatedItem

relatedItem_title :: Lens' RelatedItem (Maybe Text)

The title of the related item.

relatedItem_identifier :: Lens' RelatedItem ItemIdentifier

Details about the related item.

RelatedItemsUpdate

relatedItemsUpdate_itemToRemove :: Lens' RelatedItemsUpdate (Maybe ItemIdentifier)

Details about the related item you're deleting.

relatedItemsUpdate_itemToAdd :: Lens' RelatedItemsUpdate (Maybe RelatedItem)

Details about the related item you're adding.

ReplicationSet

replicationSet_arn :: Lens' ReplicationSet (Maybe Text)

The Amazon Resource Name (ARN) of the replication set.

replicationSet_createdBy :: Lens' ReplicationSet Text

Details about who created the replication set.

replicationSet_createdTime :: Lens' ReplicationSet UTCTime

When the replication set was created.

replicationSet_deletionProtected :: Lens' ReplicationSet Bool

Determines if the replication set deletion protection is enabled or not. If deletion protection is enabled, you can't delete the last Region in the replication set.

replicationSet_lastModifiedBy :: Lens' ReplicationSet Text

Who last modified the replication set.

replicationSet_lastModifiedTime :: Lens' ReplicationSet UTCTime

When the replication set was last updated.

replicationSet_regionMap :: Lens' ReplicationSet (HashMap Text RegionInfo)

The map between each Region in your replication set and the KMS key that is used to encrypt the data in that Region.

replicationSet_status :: Lens' ReplicationSet ReplicationSetStatus

The status of the replication set. If the replication set is still pending, you can't use Incident Manager functionality.

ResourcePolicy

resourcePolicy_policyDocument :: Lens' ResourcePolicy Text

The JSON blob that describes the policy.

resourcePolicy_policyId :: Lens' ResourcePolicy Text

The ID of the resource policy.

resourcePolicy_ramResourceShareRegion :: Lens' ResourcePolicy Text

The Region that policy allows resources to be used in.

ResponsePlanSummary

responsePlanSummary_displayName :: Lens' ResponsePlanSummary (Maybe Text)

The human readable name of the response plan. This can include spaces.

responsePlanSummary_arn :: Lens' ResponsePlanSummary Text

The Amazon Resource Name (ARN) of the response plan.

responsePlanSummary_name :: Lens' ResponsePlanSummary Text

The name of the response plan. This can't include spaces.

SsmAutomation

ssmAutomation_targetAccount :: Lens' SsmAutomation (Maybe SsmTargetAccount)

The account that the automation document will be run in. This can be in either the management account or an application account.

ssmAutomation_parameters :: Lens' SsmAutomation (Maybe (HashMap Text [Text]))

The key-value pair parameters to use when running the automation document.

ssmAutomation_documentVersion :: Lens' SsmAutomation (Maybe Text)

The automation document's version to use when running.

ssmAutomation_documentName :: Lens' SsmAutomation Text

The automation document's name.

ssmAutomation_roleArn :: Lens' SsmAutomation Text

The Amazon Resource Name (ARN) of the role that the automation document will assume when running commands.

TimelineEvent

timelineEvent_eventData :: Lens' TimelineEvent Text

A short description of the event.

timelineEvent_eventId :: Lens' TimelineEvent Text

The ID of the timeline event.

timelineEvent_eventTime :: Lens' TimelineEvent UTCTime

The time that the event occurred.

timelineEvent_eventType :: Lens' TimelineEvent Text

The type of event that occurred. Currently Incident Manager supports only the Custom Event type.

timelineEvent_eventUpdatedTime :: Lens' TimelineEvent UTCTime

The time that the timeline event was last updated.

timelineEvent_incidentRecordArn :: Lens' TimelineEvent Text

The Amazon Resource Name (ARN) of the incident that the event occurred during.

TriggerDetails

triggerDetails_rawData :: Lens' TriggerDetails (Maybe Text)

Raw data passed from either EventBridge, CloudWatch, or Incident Manager when an incident is created.

triggerDetails_triggerArn :: Lens' TriggerDetails (Maybe Text)

The ARN of the source that detected the incident.

triggerDetails_source :: Lens' TriggerDetails Text

Identifies the service that sourced the event. All events sourced from within AWS begin with "aws." Customer-generated events can have any value here, as long as it doesn't begin with "aws." We recommend the use of Java package-name style reverse domain-name strings.

triggerDetails_timestamp :: Lens' TriggerDetails UTCTime

The time that the incident was detected.

UpdateReplicationSetAction

updateReplicationSetAction_addRegionAction :: Lens' UpdateReplicationSetAction (Maybe AddRegionAction)

Details about the Region that you're adding to the replication set.

updateReplicationSetAction_deleteRegionAction :: Lens' UpdateReplicationSetAction (Maybe DeleteRegionAction)

Details about the Region that you're deleting to the replication set.