Network.AWS.SSMIncidents.StartIncident

Description

Used to start an incident from CloudWatch alarms, EventBridge events, or manually.

Creating a Request

data StartIncident

See: newStartIncident smart constructor.

Constructors

StartIncident'

Fields

clientToken :: Maybe Text

A token ensuring that the action is called only once with the specified details.

triggerDetails :: Maybe TriggerDetails

Details of what created the incident record in Incident Manager.

relatedItems :: Maybe [RelatedItem]

Add related items to the incident for other responders to use. Related items are AWS resources, external links, or files uploaded to an S3 bucket.

impact :: Maybe Natural

Defines the impact to the customers. Providing an impact overwrites the impact provided by a response plan.

Possible impacts:

1 - Critical impact, this typically relates to full application failure that impacts many to all customers.

2 - High impact, partial application failure with impact to many customers.

3 - Medium impact, the application is providing reduced service to customers.

4 - Low impact, customer might aren't impacted by the problem yet.

5 - No impact, customers aren't currently impacted but urgent action is needed to avoid impact.

title :: Maybe Text

Provide a title for the incident. Providing a title overwrites the title provided by the response plan.

responsePlanArn :: Text

The Amazon Resource Name (ARN) of the response plan that pre-defines summary, chat channels, SNS topics, runbooks, title, and impact of the incident.

newStartIncident

Arguments

:: Text
$sel:responsePlanArn:StartIncident'

-> StartIncident

Create a value of StartIncident with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:StartIncident', startIncident_clientToken - A token ensuring that the action is called only once with the specified details.

$sel:triggerDetails:StartIncident', startIncident_triggerDetails - Details of what created the incident record in Incident Manager.

$sel:relatedItems:StartIncident', startIncident_relatedItems - Add related items to the incident for other responders to use. Related items are AWS resources, external links, or files uploaded to an S3 bucket.

$sel:impact:StartIncident', startIncident_impact - Defines the impact to the customers. Providing an impact overwrites the impact provided by a response plan.

Possible impacts:

1 - Critical impact, this typically relates to full application failure that impacts many to all customers.

2 - High impact, partial application failure with impact to many customers.

3 - Medium impact, the application is providing reduced service to customers.

4 - Low impact, customer might aren't impacted by the problem yet.

5 - No impact, customers aren't currently impacted but urgent action is needed to avoid impact.

$sel:title:StartIncident', startIncident_title - Provide a title for the incident. Providing a title overwrites the title provided by the response plan.

$sel:responsePlanArn:StartIncident', startIncident_responsePlanArn - The Amazon Resource Name (ARN) of the response plan that pre-defines summary, chat channels, SNS topics, runbooks, title, and impact of the incident.

Request Lenses

startIncident_clientToken :: Lens' StartIncident (Maybe Text)

A token ensuring that the action is called only once with the specified details.

startIncident_triggerDetails :: Lens' StartIncident (Maybe TriggerDetails)

Details of what created the incident record in Incident Manager.

startIncident_relatedItems :: Lens' StartIncident (Maybe [RelatedItem])

Add related items to the incident for other responders to use. Related items are AWS resources, external links, or files uploaded to an S3 bucket.

startIncident_impact :: Lens' StartIncident (Maybe Natural)

Defines the impact to the customers. Providing an impact overwrites the impact provided by a response plan.

Possible impacts:

1 - Critical impact, this typically relates to full application failure that impacts many to all customers.

2 - High impact, partial application failure with impact to many customers.

3 - Medium impact, the application is providing reduced service to customers.

4 - Low impact, customer might aren't impacted by the problem yet.

5 - No impact, customers aren't currently impacted but urgent action is needed to avoid impact.

startIncident_title :: Lens' StartIncident (Maybe Text)

Provide a title for the incident. Providing a title overwrites the title provided by the response plan.

startIncident_responsePlanArn :: Lens' StartIncident Text

The Amazon Resource Name (ARN) of the response plan that pre-defines summary, chat channels, SNS topics, runbooks, title, and impact of the incident.

Destructuring the Response

data StartIncidentResponse

See: newStartIncidentResponse smart constructor.

Constructors

StartIncidentResponse'

Fields

httpStatus :: Int

The response's http status code.

incidentRecordArn :: Text

The ARN of the newly created incident record.

newStartIncidentResponse

Arguments

:: Int
$sel:httpStatus:StartIncidentResponse'

-> Text
$sel:incidentRecordArn:StartIncidentResponse'

-> StartIncidentResponse

Create a value of StartIncidentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StartIncidentResponse', startIncidentResponse_httpStatus - The response's http status code.

$sel:incidentRecordArn:StartIncidentResponse', startIncidentResponse_incidentRecordArn - The ARN of the newly created incident record.

Response Lenses

startIncidentResponse_httpStatus :: Lens' StartIncidentResponse Int

The response's http status code.

startIncidentResponse_incidentRecordArn :: Lens' StartIncidentResponse Text

The ARN of the newly created incident record.