Network.AWS.SSMIncidents.Types

Description

Service Configuration

defaultService :: Service

API version 2018-05-10 of the Amazon Systems Manager Incident Manager SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by an AWS service.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You don't have sufficient access to perform this action.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Updating or deleting a resource causes an inconsistent state.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Request would cause a service quota to be exceeded.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The request processing has failed because of an unknown error, exception or failure.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Request references a resource which does not exist.

IncidentRecordStatus

newtype IncidentRecordStatus

Constructors

IncidentRecordStatus'

Fields

fromIncidentRecordStatus :: Text

Bundled Patterns

pattern IncidentRecordStatus_OPEN :: IncidentRecordStatus

pattern IncidentRecordStatus_RESOLVED :: IncidentRecordStatus

ItemType

newtype ItemType

Constructors

ItemType'

Fields

fromItemType :: Text

Bundled Patterns

pattern ItemType_ANALYSIS :: ItemType

pattern ItemType_ATTACHMENT :: ItemType

pattern ItemType_INCIDENT :: ItemType

pattern ItemType_METRIC :: ItemType

pattern ItemType_OTHER :: ItemType

pattern ItemType_PARENT :: ItemType

RegionStatus

newtype RegionStatus

Constructors

RegionStatus'

Fields

fromRegionStatus :: Text

Bundled Patterns

pattern RegionStatus_ACTIVE :: RegionStatus

pattern RegionStatus_CREATING :: RegionStatus

pattern RegionStatus_DELETING :: RegionStatus

pattern RegionStatus_FAILED :: RegionStatus

ReplicationSetStatus

newtype ReplicationSetStatus

Constructors

ReplicationSetStatus'

Fields

fromReplicationSetStatus :: Text

Bundled Patterns

pattern ReplicationSetStatus_ACTIVE :: ReplicationSetStatus

pattern ReplicationSetStatus_CREATING :: ReplicationSetStatus

pattern ReplicationSetStatus_DELETING :: ReplicationSetStatus

pattern ReplicationSetStatus_FAILED :: ReplicationSetStatus

pattern ReplicationSetStatus_UPDATING :: ReplicationSetStatus

SortOrder

newtype SortOrder

Constructors

SortOrder'

Fields

fromSortOrder :: Text

Bundled Patterns

pattern SortOrder_ASCENDING :: SortOrder

pattern SortOrder_DESCENDING :: SortOrder

SsmTargetAccount

newtype SsmTargetAccount

Constructors

SsmTargetAccount'

Fields

fromSsmTargetAccount :: Text

Bundled Patterns

pattern SsmTargetAccount_IMPACTED_ACCOUNT :: SsmTargetAccount

pattern SsmTargetAccount_RESPONSE_PLAN_OWNER_ACCOUNT :: SsmTargetAccount

TimelineEventSort

newtype TimelineEventSort

Constructors

TimelineEventSort'

Fields

fromTimelineEventSort :: Text

Bundled Patterns

pattern TimelineEventSort_EVENT_TIME :: TimelineEventSort

Action

data Action

The action that starts at the beginning of an incident. The response plan defines the action.

See: newAction smart constructor.

Constructors

Action'

Fields

ssmAutomation :: Maybe SsmAutomation

The Systems Manager automation document to start as the runbook at the beginning of the incident.

newAction :: Action

Create a value of Action with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ssmAutomation:Action', action_ssmAutomation - The Systems Manager automation document to start as the runbook at the beginning of the incident.

action_ssmAutomation :: Lens' Action (Maybe SsmAutomation)

The Systems Manager automation document to start as the runbook at the beginning of the incident.

AddRegionAction

data AddRegionAction

Defines the Region and KMS key to add to the replication set.

See: newAddRegionAction smart constructor.

Constructors

AddRegionAction'

Fields

sseKmsKeyId :: Maybe Text

The KMS key ID to use to encrypt your replication set.

regionName :: Text

The Region name to add to the replication set.

newAddRegionAction

Arguments

:: Text
$sel:regionName:AddRegionAction'

-> AddRegionAction

Create a value of AddRegionAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sseKmsKeyId:AddRegionAction', addRegionAction_sseKmsKeyId - The KMS key ID to use to encrypt your replication set.

$sel:regionName:AddRegionAction', addRegionAction_regionName - The Region name to add to the replication set.

addRegionAction_sseKmsKeyId :: Lens' AddRegionAction (Maybe Text)

The KMS key ID to use to encrypt your replication set.

addRegionAction_regionName :: Lens' AddRegionAction Text

The Region name to add to the replication set.

AttributeValueList

data AttributeValueList

Use the AttributeValueList to filter by string or integer values.

See: newAttributeValueList smart constructor.

Constructors

AttributeValueList'

Fields

stringValues :: Maybe [Text]

The list of string values that the filter matches.

integerValues :: Maybe [Int]

The list of integer values that the filter matches.

newAttributeValueList :: AttributeValueList

Create a value of AttributeValueList with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stringValues:AttributeValueList', attributeValueList_stringValues - The list of string values that the filter matches.

$sel:integerValues:AttributeValueList', attributeValueList_integerValues - The list of integer values that the filter matches.

attributeValueList_stringValues :: Lens' AttributeValueList (Maybe [Text])

The list of string values that the filter matches.

attributeValueList_integerValues :: Lens' AttributeValueList (Maybe [Int])

The list of integer values that the filter matches.

AutomationExecution

data AutomationExecution

The Systems Manager automation document process to start as the runbook at the beginning of the incident.

See: newAutomationExecution smart constructor.

Constructors

AutomationExecution'

Fields

ssmExecutionArn :: Maybe Text

The Amazon Resource Name (ARN) of the automation process.

newAutomationExecution :: AutomationExecution

Create a value of AutomationExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ssmExecutionArn:AutomationExecution', automationExecution_ssmExecutionArn - The Amazon Resource Name (ARN) of the automation process.

automationExecution_ssmExecutionArn :: Lens' AutomationExecution (Maybe Text)

The Amazon Resource Name (ARN) of the automation process.

ChatChannel

data ChatChannel

The AWS Chatbot chat channel used for collaboration during an incident.

See: newChatChannel smart constructor.

Constructors

ChatChannel'

Fields

empty :: Maybe EmptyChatChannel

Used to remove the chat channel from an incident record or response plan.

chatbotSns :: Maybe (NonEmpty Text)

The SNS targets that AWS Chatbot uses to notify the chat channel of updates to an incident. You can also make updates to the incident through the chat channel by using the SNS topics.

newChatChannel :: ChatChannel

Create a value of ChatChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:empty:ChatChannel', chatChannel_empty - Used to remove the chat channel from an incident record or response plan.

$sel:chatbotSns:ChatChannel', chatChannel_chatbotSns - The SNS targets that AWS Chatbot uses to notify the chat channel of updates to an incident. You can also make updates to the incident through the chat channel by using the SNS topics.

chatChannel_empty :: Lens' ChatChannel (Maybe EmptyChatChannel)

Used to remove the chat channel from an incident record or response plan.

chatChannel_chatbotSns :: Lens' ChatChannel (Maybe (NonEmpty Text))

The SNS targets that AWS Chatbot uses to notify the chat channel of updates to an incident. You can also make updates to the incident through the chat channel by using the SNS topics.

Condition

data Condition

A conditional statement with which to compare a value, after a timestamp, before a timestamp, or equal to a string or integer. If multiple conditions are specified, the conditionals become an ANDed statement. If multiple values are specified for a conditional, the values are ORd.

See: newCondition smart constructor.

Constructors

Condition'

Fields

after :: Maybe POSIX

After the specified timestamp.

equals :: Maybe AttributeValueList

The value is equal to the provided string or integer.

before :: Maybe POSIX

Before the specified timestamp

newCondition :: Condition

Create a value of Condition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:after:Condition', condition_after - After the specified timestamp.

$sel:equals:Condition', condition_equals - The value is equal to the provided string or integer.

$sel:before:Condition', condition_before - Before the specified timestamp

condition_after :: Lens' Condition (Maybe UTCTime)

After the specified timestamp.

condition_equals :: Lens' Condition (Maybe AttributeValueList)

The value is equal to the provided string or integer.

condition_before :: Lens' Condition (Maybe UTCTime)

Before the specified timestamp

DeleteRegionAction

data DeleteRegionAction

Defines the information about the Region you're deleting from your replication set.

See: newDeleteRegionAction smart constructor.

Constructors

DeleteRegionAction'

Fields

regionName :: Text

The name of the Region you're deleting from the replication set.

newDeleteRegionAction

Arguments

:: Text
$sel:regionName:DeleteRegionAction'

-> DeleteRegionAction

Create a value of DeleteRegionAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:regionName:DeleteRegionAction', deleteRegionAction_regionName - The name of the Region you're deleting from the replication set.

deleteRegionAction_regionName :: Lens' DeleteRegionAction Text

The name of the Region you're deleting from the replication set.

EmptyChatChannel

data EmptyChatChannel

Used to remove the chat channel from an incident record or response plan.

See: newEmptyChatChannel smart constructor.

Constructors

EmptyChatChannel'

newEmptyChatChannel :: EmptyChatChannel

Create a value of EmptyChatChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

EventSummary

data EventSummary

Details about a timeline event during an incident.

See: newEventSummary smart constructor.

Constructors

EventSummary'

Fields

eventId :: Text

The timeline event ID.

eventTime :: POSIX

The time that the event occurred.

eventType :: Text

The type of event. The timeline event must be Custom Event.

eventUpdatedTime :: POSIX

The time that the timeline event was last updated.

incidentRecordArn :: Text

The Amazon Resource Name (ARN) of the incident that the event happened during.

newEventSummary

Arguments

:: Text
$sel:eventId:EventSummary'

-> UTCTime
$sel:eventTime:EventSummary'

-> Text
$sel:eventType:EventSummary'

-> UTCTime
$sel:eventUpdatedTime:EventSummary'

-> Text
$sel:incidentRecordArn:EventSummary'

-> EventSummary

Create a value of EventSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventId:EventSummary', eventSummary_eventId - The timeline event ID.

$sel:eventTime:EventSummary', eventSummary_eventTime - The time that the event occurred.

$sel:eventType:EventSummary', eventSummary_eventType - The type of event. The timeline event must be Custom Event.

$sel:eventUpdatedTime:EventSummary', eventSummary_eventUpdatedTime - The time that the timeline event was last updated.

$sel:incidentRecordArn:EventSummary', eventSummary_incidentRecordArn - The Amazon Resource Name (ARN) of the incident that the event happened during.

eventSummary_eventId :: Lens' EventSummary Text

The timeline event ID.

eventSummary_eventTime :: Lens' EventSummary UTCTime

The time that the event occurred.

eventSummary_eventType :: Lens' EventSummary Text

The type of event. The timeline event must be Custom Event.

eventSummary_eventUpdatedTime :: Lens' EventSummary UTCTime

The time that the timeline event was last updated.

eventSummary_incidentRecordArn :: Lens' EventSummary Text

The Amazon Resource Name (ARN) of the incident that the event happened during.

Filter

data Filter

Filter the selection by using a condition.

See: newFilter smart constructor.

Constructors

Filter'

Fields

condition :: Condition

The condition accepts before or after a specified time, equal to a string, or equal to an integer.

key :: Text

The key that you're filtering on.

newFilter

Arguments

:: Condition
$sel:condition:Filter'

-> Text
$sel:key:Filter'

-> Filter

Create a value of Filter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:condition:Filter', filter_condition - The condition accepts before or after a specified time, equal to a string, or equal to an integer.

$sel:key:Filter', filter_key - The key that you're filtering on.

filter_condition :: Lens' Filter Condition

The condition accepts before or after a specified time, equal to a string, or equal to an integer.

filter_key :: Lens' Filter Text

The key that you're filtering on.

IncidentRecord

data IncidentRecord

The record of the incident that's created when an incident occurs.

See: newIncidentRecord smart constructor.

Constructors

IncidentRecord'

Fields

summary :: Maybe Text

The summary of the incident. The summary is a brief synopsis of what occurred, what is currently happening, and context.

notificationTargets :: Maybe [NotificationTargetItem]

The SNS targets that are notified when updates are made to an incident.

resolvedTime :: Maybe POSIX

The time at which the incident was resolved. This appears as a timeline event.

chatChannel :: Maybe ChatChannel

The chat channel used for collaboration during an incident.

automationExecutions :: Maybe [AutomationExecution]

The runbook, or automation document, that's run at the beginning of the incident.

arn :: Text

The Amazon Resource Name (ARN) of the incident record.

creationTime :: POSIX

The time that Incident Manager created the incident record.

dedupeString :: Text

The string Incident Manager uses to prevent duplicate incidents from being created by the same incident.

impact :: Natural

The impact of the incident on customers and applications.

incidentRecordSource :: IncidentRecordSource

Details about the action that started the incident.

lastModifiedBy :: Text

Who modified the incident most recently.

lastModifiedTime :: POSIX

The time at which the incident was most recently modified.

status :: IncidentRecordStatus

The current status of the incident.

title :: Text

The title of the incident.

newIncidentRecord

Arguments

:: Text
$sel:arn:IncidentRecord'

-> UTCTime
$sel:creationTime:IncidentRecord'

-> Text
$sel:dedupeString:IncidentRecord'

-> Natural
$sel:impact:IncidentRecord'

-> IncidentRecordSource
$sel:incidentRecordSource:IncidentRecord'

-> Text
$sel:lastModifiedBy:IncidentRecord'

-> UTCTime
$sel:lastModifiedTime:IncidentRecord'

-> IncidentRecordStatus
$sel:status:IncidentRecord'

-> Text
$sel:title:IncidentRecord'

-> IncidentRecord

Create a value of IncidentRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:IncidentRecord', incidentRecord_summary - The summary of the incident. The summary is a brief synopsis of what occurred, what is currently happening, and context.

$sel:notificationTargets:IncidentRecord', incidentRecord_notificationTargets - The SNS targets that are notified when updates are made to an incident.

$sel:resolvedTime:IncidentRecord', incidentRecord_resolvedTime - The time at which the incident was resolved. This appears as a timeline event.

$sel:chatChannel:IncidentRecord', incidentRecord_chatChannel - The chat channel used for collaboration during an incident.

$sel:automationExecutions:IncidentRecord', incidentRecord_automationExecutions - The runbook, or automation document, that's run at the beginning of the incident.

$sel:arn:IncidentRecord', incidentRecord_arn - The Amazon Resource Name (ARN) of the incident record.

$sel:creationTime:IncidentRecord', incidentRecord_creationTime - The time that Incident Manager created the incident record.

$sel:dedupeString:IncidentRecord', incidentRecord_dedupeString - The string Incident Manager uses to prevent duplicate incidents from being created by the same incident.

$sel:impact:IncidentRecord', incidentRecord_impact - The impact of the incident on customers and applications.

$sel:incidentRecordSource:IncidentRecord', incidentRecord_incidentRecordSource - Details about the action that started the incident.

$sel:lastModifiedBy:IncidentRecord', incidentRecord_lastModifiedBy - Who modified the incident most recently.

$sel:lastModifiedTime:IncidentRecord', incidentRecord_lastModifiedTime - The time at which the incident was most recently modified.

$sel:status:IncidentRecord', incidentRecord_status - The current status of the incident.

$sel:title:IncidentRecord', incidentRecord_title - The title of the incident.

incidentRecord_summary :: Lens' IncidentRecord (Maybe Text)

The summary of the incident. The summary is a brief synopsis of what occurred, what is currently happening, and context.

incidentRecord_notificationTargets :: Lens' IncidentRecord (Maybe [NotificationTargetItem])

The SNS targets that are notified when updates are made to an incident.

incidentRecord_resolvedTime :: Lens' IncidentRecord (Maybe UTCTime)

The time at which the incident was resolved. This appears as a timeline event.

incidentRecord_chatChannel :: Lens' IncidentRecord (Maybe ChatChannel)

The chat channel used for collaboration during an incident.

incidentRecord_automationExecutions :: Lens' IncidentRecord (Maybe [AutomationExecution])

The runbook, or automation document, that's run at the beginning of the incident.

incidentRecord_arn :: Lens' IncidentRecord Text

The Amazon Resource Name (ARN) of the incident record.

incidentRecord_creationTime :: Lens' IncidentRecord UTCTime

The time that Incident Manager created the incident record.

incidentRecord_dedupeString :: Lens' IncidentRecord Text

The string Incident Manager uses to prevent duplicate incidents from being created by the same incident.

incidentRecord_impact :: Lens' IncidentRecord Natural

The impact of the incident on customers and applications.

incidentRecord_incidentRecordSource :: Lens' IncidentRecord IncidentRecordSource

Details about the action that started the incident.

incidentRecord_lastModifiedBy :: Lens' IncidentRecord Text

Who modified the incident most recently.

incidentRecord_lastModifiedTime :: Lens' IncidentRecord UTCTime

The time at which the incident was most recently modified.

incidentRecord_status :: Lens' IncidentRecord IncidentRecordStatus

The current status of the incident.

incidentRecord_title :: Lens' IncidentRecord Text

The title of the incident.

IncidentRecordSource

data IncidentRecordSource

Details about how the incident record was created and when.

See: newIncidentRecordSource smart constructor.

Constructors

IncidentRecordSource'

Fields

invokedBy :: Maybe Text

The principal the assumed the role specified of the createdBy.

resourceArn :: Maybe Text

The resource that caused the incident to be created.

createdBy :: Text

The principal that started the incident.

source :: Text

The service that started the incident. This can be manually created from Incident Manager, automatically created using an AWS CloudWatch alarm, or Amazon EventBridge event.

newIncidentRecordSource

Arguments

:: Text
$sel:createdBy:IncidentRecordSource'

-> Text
$sel:source:IncidentRecordSource'

-> IncidentRecordSource

Create a value of IncidentRecordSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:invokedBy:IncidentRecordSource', incidentRecordSource_invokedBy - The principal the assumed the role specified of the createdBy.

$sel:resourceArn:IncidentRecordSource', incidentRecordSource_resourceArn - The resource that caused the incident to be created.

$sel:createdBy:IncidentRecordSource', incidentRecordSource_createdBy - The principal that started the incident.

$sel:source:IncidentRecordSource', incidentRecordSource_source - The service that started the incident. This can be manually created from Incident Manager, automatically created using an AWS CloudWatch alarm, or Amazon EventBridge event.

incidentRecordSource_invokedBy :: Lens' IncidentRecordSource (Maybe Text)

The principal the assumed the role specified of the createdBy.

incidentRecordSource_resourceArn :: Lens' IncidentRecordSource (Maybe Text)

The resource that caused the incident to be created.

incidentRecordSource_createdBy :: Lens' IncidentRecordSource Text

The principal that started the incident.

incidentRecordSource_source :: Lens' IncidentRecordSource Text

The service that started the incident. This can be manually created from Incident Manager, automatically created using an AWS CloudWatch alarm, or Amazon EventBridge event.

IncidentRecordSummary

data IncidentRecordSummary

Details describing an incident record.

See: newIncidentRecordSummary smart constructor.

Constructors

IncidentRecordSummary'

Fields

resolvedTime :: Maybe POSIX

The time the incident was resolved.

arn :: Text

The Amazon Resource Name (ARN) of the incident.

creationTime :: POSIX

The time the incident was created.

impact :: Natural

Defines the impact to customers and applications.

incidentRecordSource :: IncidentRecordSource

What caused Incident Manager to create the incident.

status :: IncidentRecordStatus

The current status of the incident.

title :: Text

The title of the incident. This value is either provided by the response plan or overwritten on creation.

newIncidentRecordSummary

Arguments

:: Text
$sel:arn:IncidentRecordSummary'

-> UTCTime
$sel:creationTime:IncidentRecordSummary'

-> Natural
$sel:impact:IncidentRecordSummary'

-> IncidentRecordSource
$sel:incidentRecordSource:IncidentRecordSummary'

-> IncidentRecordStatus
$sel:status:IncidentRecordSummary'

-> Text
$sel:title:IncidentRecordSummary'

-> IncidentRecordSummary

Create a value of IncidentRecordSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolvedTime:IncidentRecordSummary', incidentRecordSummary_resolvedTime - The time the incident was resolved.

$sel:arn:IncidentRecordSummary', incidentRecordSummary_arn - The Amazon Resource Name (ARN) of the incident.

$sel:creationTime:IncidentRecordSummary', incidentRecordSummary_creationTime - The time the incident was created.

$sel:impact:IncidentRecordSummary', incidentRecordSummary_impact - Defines the impact to customers and applications.

$sel:incidentRecordSource:IncidentRecordSummary', incidentRecordSummary_incidentRecordSource - What caused Incident Manager to create the incident.

$sel:status:IncidentRecordSummary', incidentRecordSummary_status - The current status of the incident.

$sel:title:IncidentRecordSummary', incidentRecordSummary_title - The title of the incident. This value is either provided by the response plan or overwritten on creation.

incidentRecordSummary_resolvedTime :: Lens' IncidentRecordSummary (Maybe UTCTime)

The time the incident was resolved.

incidentRecordSummary_arn :: Lens' IncidentRecordSummary Text

The Amazon Resource Name (ARN) of the incident.

incidentRecordSummary_creationTime :: Lens' IncidentRecordSummary UTCTime

The time the incident was created.

incidentRecordSummary_impact :: Lens' IncidentRecordSummary Natural

Defines the impact to customers and applications.

incidentRecordSummary_incidentRecordSource :: Lens' IncidentRecordSummary IncidentRecordSource

What caused Incident Manager to create the incident.

incidentRecordSummary_status :: Lens' IncidentRecordSummary IncidentRecordStatus

The current status of the incident.

incidentRecordSummary_title :: Lens' IncidentRecordSummary Text

The title of the incident. This value is either provided by the response plan or overwritten on creation.

IncidentTemplate

data IncidentTemplate

Basic details used in creating a response plan. The response plan is then used to create an incident record.

See: newIncidentTemplate smart constructor.

Constructors

IncidentTemplate'

Fields

summary :: Maybe Text

The summary of the incident. The summary is a brief synopsis of what occurred, what's currently happening, and context.

notificationTargets :: Maybe [NotificationTargetItem]

The SNS targets that are notified when updates are made to an incident.

dedupeString :: Maybe Text

Used to stop Incident Manager from creating multiple incident records for the same incident.

impact :: Natural

The impact of the incident on your customers and applications.

title :: Text

The title of the incident.

newIncidentTemplate

Arguments

:: Natural
$sel:impact:IncidentTemplate'

-> Text
$sel:title:IncidentTemplate'

-> IncidentTemplate

Create a value of IncidentTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:IncidentTemplate', incidentTemplate_summary - The summary of the incident. The summary is a brief synopsis of what occurred, what's currently happening, and context.

$sel:notificationTargets:IncidentTemplate', incidentTemplate_notificationTargets - The SNS targets that are notified when updates are made to an incident.

$sel:dedupeString:IncidentTemplate', incidentTemplate_dedupeString - Used to stop Incident Manager from creating multiple incident records for the same incident.

$sel:impact:IncidentTemplate', incidentTemplate_impact - The impact of the incident on your customers and applications.

$sel:title:IncidentTemplate', incidentTemplate_title - The title of the incident.

incidentTemplate_summary :: Lens' IncidentTemplate (Maybe Text)

The summary of the incident. The summary is a brief synopsis of what occurred, what's currently happening, and context.

incidentTemplate_notificationTargets :: Lens' IncidentTemplate (Maybe [NotificationTargetItem])

The SNS targets that are notified when updates are made to an incident.

incidentTemplate_dedupeString :: Lens' IncidentTemplate (Maybe Text)

Used to stop Incident Manager from creating multiple incident records for the same incident.

incidentTemplate_impact :: Lens' IncidentTemplate Natural

The impact of the incident on your customers and applications.

incidentTemplate_title :: Lens' IncidentTemplate Text

The title of the incident.

ItemIdentifier

data ItemIdentifier

Details and type of a related item.

See: newItemIdentifier smart constructor.

Constructors

ItemIdentifier'

Fields

type' :: ItemType

The type of related item. Incident Manager supports the following types:

ANALYSIS

INCIDENT

METRIC

PARENT

ATTACHMENT

OTHER

value :: ItemValue

Details about the related item.

newItemIdentifier

Arguments

:: ItemType
$sel:type':ItemIdentifier'

-> ItemValue
$sel:value:ItemIdentifier'

-> ItemIdentifier

Create a value of ItemIdentifier with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':ItemIdentifier', itemIdentifier_type - The type of related item. Incident Manager supports the following types:

ANALYSIS

INCIDENT

METRIC

PARENT

ATTACHMENT

OTHER

$sel:value:ItemIdentifier', itemIdentifier_value - Details about the related item.

itemIdentifier_type :: Lens' ItemIdentifier ItemType

The type of related item. Incident Manager supports the following types:

ANALYSIS

INCIDENT

METRIC

PARENT

ATTACHMENT

OTHER

itemIdentifier_value :: Lens' ItemIdentifier ItemValue

Details about the related item.

ItemValue

data ItemValue

Describes a related item.

See: newItemValue smart constructor.

Constructors

ItemValue'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the related item, if the related item is an Amazon resource.

url :: Maybe Text

The URL, if the related item is a non-AWS resource.

metricDefinition :: Maybe Text

The metric definition, if the related item is a metric in CloudWatch.

newItemValue :: ItemValue

Create a value of ItemValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ItemValue', itemValue_arn - The Amazon Resource Name (ARN) of the related item, if the related item is an Amazon resource.

$sel:url:ItemValue', itemValue_url - The URL, if the related item is a non-AWS resource.

$sel:metricDefinition:ItemValue', itemValue_metricDefinition - The metric definition, if the related item is a metric in CloudWatch.

itemValue_arn :: Lens' ItemValue (Maybe Text)

The Amazon Resource Name (ARN) of the related item, if the related item is an Amazon resource.

itemValue_url :: Lens' ItemValue (Maybe Text)

The URL, if the related item is a non-AWS resource.

itemValue_metricDefinition :: Lens' ItemValue (Maybe Text)

The metric definition, if the related item is a metric in CloudWatch.

NotificationTargetItem

data NotificationTargetItem

The SNS targets that are notified when updates are made to an incident.

See: newNotificationTargetItem smart constructor.

Constructors

NotificationTargetItem'

Fields

snsTopicArn :: Maybe Text

The Amazon Resource Name (ARN) of the SNS topic.

newNotificationTargetItem :: NotificationTargetItem

Create a value of NotificationTargetItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snsTopicArn:NotificationTargetItem', notificationTargetItem_snsTopicArn - The Amazon Resource Name (ARN) of the SNS topic.

notificationTargetItem_snsTopicArn :: Lens' NotificationTargetItem (Maybe Text)

The Amazon Resource Name (ARN) of the SNS topic.

RegionInfo

data RegionInfo

Information about a Region in your replication set.

See: newRegionInfo smart constructor.

Constructors

RegionInfo'

Fields

statusMessage :: Maybe Text

Information displayed about the status of the Region.

sseKmsKeyId :: Maybe Text

The ID of the KMS key used to encrypt the data in this Region.

status :: RegionStatus

The status of the Region in the replication set.

statusUpdateDateTime :: POSIX

The most recent date and time that the Region's status was updated.

newRegionInfo

Arguments

:: RegionStatus
$sel:status:RegionInfo'

-> UTCTime
$sel:statusUpdateDateTime:RegionInfo'

-> RegionInfo

Create a value of RegionInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusMessage:RegionInfo', regionInfo_statusMessage - Information displayed about the status of the Region.

$sel:sseKmsKeyId:RegionInfo', regionInfo_sseKmsKeyId - The ID of the KMS key used to encrypt the data in this Region.

$sel:status:RegionInfo', regionInfo_status - The status of the Region in the replication set.

$sel:statusUpdateDateTime:RegionInfo', regionInfo_statusUpdateDateTime - The most recent date and time that the Region's status was updated.

regionInfo_statusMessage :: Lens' RegionInfo (Maybe Text)

Information displayed about the status of the Region.

regionInfo_sseKmsKeyId :: Lens' RegionInfo (Maybe Text)

The ID of the KMS key used to encrypt the data in this Region.

regionInfo_status :: Lens' RegionInfo RegionStatus

The status of the Region in the replication set.

regionInfo_statusUpdateDateTime :: Lens' RegionInfo UTCTime

The most recent date and time that the Region's status was updated.

RegionMapInputValue

data RegionMapInputValue

The mapping between a Region and the key that's used to encrypt the data.

See: newRegionMapInputValue smart constructor.

Constructors

RegionMapInputValue'

Fields

sseKmsKeyId :: Maybe Text

The KMS key used to encrypt the data in your replication set.

newRegionMapInputValue :: RegionMapInputValue

Create a value of RegionMapInputValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sseKmsKeyId:RegionMapInputValue', regionMapInputValue_sseKmsKeyId - The KMS key used to encrypt the data in your replication set.

regionMapInputValue_sseKmsKeyId :: Lens' RegionMapInputValue (Maybe Text)

The KMS key used to encrypt the data in your replication set.

RelatedItem

data RelatedItem

Resources that responders use to triage and mitigate the incident.

See: newRelatedItem smart constructor.

Constructors

RelatedItem'

Fields

title :: Maybe Text

The title of the related item.

identifier :: ItemIdentifier

Details about the related item.

newRelatedItem

Arguments

:: ItemIdentifier
$sel:identifier:RelatedItem'

-> RelatedItem

Create a value of RelatedItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:title:RelatedItem', relatedItem_title - The title of the related item.

$sel:identifier:RelatedItem', relatedItem_identifier - Details about the related item.

relatedItem_title :: Lens' RelatedItem (Maybe Text)

The title of the related item.

relatedItem_identifier :: Lens' RelatedItem ItemIdentifier

Details about the related item.

RelatedItemsUpdate

data RelatedItemsUpdate

Details about the related item you're adding.

See: newRelatedItemsUpdate smart constructor.

Constructors

RelatedItemsUpdate'

Fields

itemToRemove :: Maybe ItemIdentifier

Details about the related item you're deleting.

itemToAdd :: Maybe RelatedItem

Details about the related item you're adding.

newRelatedItemsUpdate :: RelatedItemsUpdate

Create a value of RelatedItemsUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:itemToRemove:RelatedItemsUpdate', relatedItemsUpdate_itemToRemove - Details about the related item you're deleting.

$sel:itemToAdd:RelatedItemsUpdate', relatedItemsUpdate_itemToAdd - Details about the related item you're adding.

relatedItemsUpdate_itemToRemove :: Lens' RelatedItemsUpdate (Maybe ItemIdentifier)

Details about the related item you're deleting.

relatedItemsUpdate_itemToAdd :: Lens' RelatedItemsUpdate (Maybe RelatedItem)

Details about the related item you're adding.

ReplicationSet

data ReplicationSet

The set of Regions that your Incident Manager data will be replicated to and the KMS key used to encrypt the data.

See: newReplicationSet smart constructor.

Constructors

ReplicationSet'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the replication set.

createdBy :: Text

Details about who created the replication set.

createdTime :: POSIX

When the replication set was created.

deletionProtected :: Bool

Determines if the replication set deletion protection is enabled or not. If deletion protection is enabled, you can't delete the last Region in the replication set.

lastModifiedBy :: Text

Who last modified the replication set.

lastModifiedTime :: POSIX

When the replication set was last updated.

regionMap :: HashMap Text RegionInfo

The map between each Region in your replication set and the KMS key that is used to encrypt the data in that Region.

status :: ReplicationSetStatus

The status of the replication set. If the replication set is still pending, you can't use Incident Manager functionality.

newReplicationSet

Arguments

:: Text
$sel:createdBy:ReplicationSet'

-> UTCTime
$sel:createdTime:ReplicationSet'

-> Bool
$sel:deletionProtected:ReplicationSet'

-> Text
$sel:lastModifiedBy:ReplicationSet'

-> UTCTime
$sel:lastModifiedTime:ReplicationSet'

-> ReplicationSetStatus
$sel:status:ReplicationSet'

-> ReplicationSet

Create a value of ReplicationSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ReplicationSet', replicationSet_arn - The Amazon Resource Name (ARN) of the replication set.

$sel:createdBy:ReplicationSet', replicationSet_createdBy - Details about who created the replication set.

$sel:createdTime:ReplicationSet', replicationSet_createdTime - When the replication set was created.

$sel:deletionProtected:ReplicationSet', replicationSet_deletionProtected - Determines if the replication set deletion protection is enabled or not. If deletion protection is enabled, you can't delete the last Region in the replication set.

$sel:lastModifiedBy:ReplicationSet', replicationSet_lastModifiedBy - Who last modified the replication set.

$sel:lastModifiedTime:ReplicationSet', replicationSet_lastModifiedTime - When the replication set was last updated.

$sel:regionMap:ReplicationSet', replicationSet_regionMap - The map between each Region in your replication set and the KMS key that is used to encrypt the data in that Region.

$sel:status:ReplicationSet', replicationSet_status - The status of the replication set. If the replication set is still pending, you can't use Incident Manager functionality.

replicationSet_arn :: Lens' ReplicationSet (Maybe Text)

The Amazon Resource Name (ARN) of the replication set.

replicationSet_createdBy :: Lens' ReplicationSet Text

Details about who created the replication set.

replicationSet_createdTime :: Lens' ReplicationSet UTCTime

When the replication set was created.

replicationSet_deletionProtected :: Lens' ReplicationSet Bool

Determines if the replication set deletion protection is enabled or not. If deletion protection is enabled, you can't delete the last Region in the replication set.

replicationSet_lastModifiedBy :: Lens' ReplicationSet Text

Who last modified the replication set.

replicationSet_lastModifiedTime :: Lens' ReplicationSet UTCTime

When the replication set was last updated.

replicationSet_regionMap :: Lens' ReplicationSet (HashMap Text RegionInfo)

The map between each Region in your replication set and the KMS key that is used to encrypt the data in that Region.

replicationSet_status :: Lens' ReplicationSet ReplicationSetStatus

The status of the replication set. If the replication set is still pending, you can't use Incident Manager functionality.

ResourcePolicy

data ResourcePolicy

The resource policy that allows Incident Manager to perform actions on resources on your behalf.

See: newResourcePolicy smart constructor.

Constructors

ResourcePolicy'

Fields

policyDocument :: Text

The JSON blob that describes the policy.

policyId :: Text

The ID of the resource policy.

ramResourceShareRegion :: Text

The Region that policy allows resources to be used in.

newResourcePolicy

Arguments

:: Text
$sel:policyDocument:ResourcePolicy'

-> Text
$sel:policyId:ResourcePolicy'

-> Text
$sel:ramResourceShareRegion:ResourcePolicy'

-> ResourcePolicy

Create a value of ResourcePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyDocument:ResourcePolicy', resourcePolicy_policyDocument - The JSON blob that describes the policy.

$sel:policyId:ResourcePolicy', resourcePolicy_policyId - The ID of the resource policy.

$sel:ramResourceShareRegion:ResourcePolicy', resourcePolicy_ramResourceShareRegion - The Region that policy allows resources to be used in.

resourcePolicy_policyDocument :: Lens' ResourcePolicy Text

The JSON blob that describes the policy.

resourcePolicy_policyId :: Lens' ResourcePolicy Text

The ID of the resource policy.

resourcePolicy_ramResourceShareRegion :: Lens' ResourcePolicy Text

The Region that policy allows resources to be used in.

ResponsePlanSummary

data ResponsePlanSummary

Details of the response plan that are used when creating an incident.

See: newResponsePlanSummary smart constructor.

Constructors

ResponsePlanSummary'

Fields

displayName :: Maybe Text

The human readable name of the response plan. This can include spaces.

arn :: Text

The Amazon Resource Name (ARN) of the response plan.

name :: Text

The name of the response plan. This can't include spaces.

newResponsePlanSummary

Arguments

:: Text
$sel:arn:ResponsePlanSummary'

-> Text
$sel:name:ResponsePlanSummary'

-> ResponsePlanSummary

Create a value of ResponsePlanSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:displayName:ResponsePlanSummary', responsePlanSummary_displayName - The human readable name of the response plan. This can include spaces.

$sel:arn:ResponsePlanSummary', responsePlanSummary_arn - The Amazon Resource Name (ARN) of the response plan.

$sel:name:ResponsePlanSummary', responsePlanSummary_name - The name of the response plan. This can't include spaces.

responsePlanSummary_displayName :: Lens' ResponsePlanSummary (Maybe Text)

The human readable name of the response plan. This can include spaces.

responsePlanSummary_arn :: Lens' ResponsePlanSummary Text

The Amazon Resource Name (ARN) of the response plan.

responsePlanSummary_name :: Lens' ResponsePlanSummary Text

The name of the response plan. This can't include spaces.

SsmAutomation

data SsmAutomation

Details about the Systems Manager automation document that will be used as a runbook during an incident.

See: newSsmAutomation smart constructor.

Constructors

SsmAutomation'

Fields

targetAccount :: Maybe SsmTargetAccount

The account that the automation document will be run in. This can be in either the management account or an application account.

parameters :: Maybe (HashMap Text [Text])

The key-value pair parameters to use when running the automation document.

documentVersion :: Maybe Text

The automation document's version to use when running.

documentName :: Text

The automation document's name.

roleArn :: Text

The Amazon Resource Name (ARN) of the role that the automation document will assume when running commands.

newSsmAutomation

Arguments

:: Text
$sel:documentName:SsmAutomation'

-> Text
$sel:roleArn:SsmAutomation'

-> SsmAutomation

Create a value of SsmAutomation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetAccount:SsmAutomation', ssmAutomation_targetAccount - The account that the automation document will be run in. This can be in either the management account or an application account.

$sel:parameters:SsmAutomation', ssmAutomation_parameters - The key-value pair parameters to use when running the automation document.

$sel:documentVersion:SsmAutomation', ssmAutomation_documentVersion - The automation document's version to use when running.

$sel:documentName:SsmAutomation', ssmAutomation_documentName - The automation document's name.

$sel:roleArn:SsmAutomation', ssmAutomation_roleArn - The Amazon Resource Name (ARN) of the role that the automation document will assume when running commands.

ssmAutomation_targetAccount :: Lens' SsmAutomation (Maybe SsmTargetAccount)

The account that the automation document will be run in. This can be in either the management account or an application account.

ssmAutomation_parameters :: Lens' SsmAutomation (Maybe (HashMap Text [Text]))

The key-value pair parameters to use when running the automation document.

ssmAutomation_documentVersion :: Lens' SsmAutomation (Maybe Text)

The automation document's version to use when running.

ssmAutomation_documentName :: Lens' SsmAutomation Text

The automation document's name.

ssmAutomation_roleArn :: Lens' SsmAutomation Text

The Amazon Resource Name (ARN) of the role that the automation document will assume when running commands.

TimelineEvent

data TimelineEvent

A significant event that happened during the incident.

See: newTimelineEvent smart constructor.

Constructors

TimelineEvent'

Fields

eventData :: Text

A short description of the event.

eventId :: Text

The ID of the timeline event.

eventTime :: POSIX

The time that the event occurred.

eventType :: Text

The type of event that occurred. Currently Incident Manager supports only the Custom Event type.

eventUpdatedTime :: POSIX

The time that the timeline event was last updated.

incidentRecordArn :: Text

The Amazon Resource Name (ARN) of the incident that the event occurred during.

newTimelineEvent

Arguments

:: Text
$sel:eventData:TimelineEvent'

-> Text
$sel:eventId:TimelineEvent'

-> UTCTime
$sel:eventTime:TimelineEvent'

-> Text
$sel:eventType:TimelineEvent'

-> UTCTime
$sel:eventUpdatedTime:TimelineEvent'

-> Text
$sel:incidentRecordArn:TimelineEvent'

-> TimelineEvent

Create a value of TimelineEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventData:TimelineEvent', timelineEvent_eventData - A short description of the event.

$sel:eventId:TimelineEvent', timelineEvent_eventId - The ID of the timeline event.

$sel:eventTime:TimelineEvent', timelineEvent_eventTime - The time that the event occurred.

$sel:eventType:TimelineEvent', timelineEvent_eventType - The type of event that occurred. Currently Incident Manager supports only the Custom Event type.

$sel:eventUpdatedTime:TimelineEvent', timelineEvent_eventUpdatedTime - The time that the timeline event was last updated.

$sel:incidentRecordArn:TimelineEvent', timelineEvent_incidentRecordArn - The Amazon Resource Name (ARN) of the incident that the event occurred during.

timelineEvent_eventData :: Lens' TimelineEvent Text

A short description of the event.

timelineEvent_eventId :: Lens' TimelineEvent Text

The ID of the timeline event.

timelineEvent_eventTime :: Lens' TimelineEvent UTCTime

The time that the event occurred.

timelineEvent_eventType :: Lens' TimelineEvent Text

The type of event that occurred. Currently Incident Manager supports only the Custom Event type.

timelineEvent_eventUpdatedTime :: Lens' TimelineEvent UTCTime

The time that the timeline event was last updated.

timelineEvent_incidentRecordArn :: Lens' TimelineEvent Text

The Amazon Resource Name (ARN) of the incident that the event occurred during.

TriggerDetails

data TriggerDetails

Details about what caused the incident to be created in Incident Manager.

See: newTriggerDetails smart constructor.

Constructors

TriggerDetails'

Fields

rawData :: Maybe Text

Raw data passed from either EventBridge, CloudWatch, or Incident Manager when an incident is created.

triggerArn :: Maybe Text

The ARN of the source that detected the incident.

source :: Text

Identifies the service that sourced the event. All events sourced from within AWS begin with "aws." Customer-generated events can have any value here, as long as it doesn't begin with "aws." We recommend the use of Java package-name style reverse domain-name strings.

timestamp :: POSIX

The time that the incident was detected.

newTriggerDetails

Arguments

:: Text
$sel:source:TriggerDetails'

-> UTCTime
$sel:timestamp:TriggerDetails'

-> TriggerDetails

Create a value of TriggerDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rawData:TriggerDetails', triggerDetails_rawData - Raw data passed from either EventBridge, CloudWatch, or Incident Manager when an incident is created.

$sel:triggerArn:TriggerDetails', triggerDetails_triggerArn - The ARN of the source that detected the incident.

$sel:source:TriggerDetails', triggerDetails_source - Identifies the service that sourced the event. All events sourced from within AWS begin with "aws." Customer-generated events can have any value here, as long as it doesn't begin with "aws." We recommend the use of Java package-name style reverse domain-name strings.

$sel:timestamp:TriggerDetails', triggerDetails_timestamp - The time that the incident was detected.

triggerDetails_rawData :: Lens' TriggerDetails (Maybe Text)

Raw data passed from either EventBridge, CloudWatch, or Incident Manager when an incident is created.

triggerDetails_triggerArn :: Lens' TriggerDetails (Maybe Text)

The ARN of the source that detected the incident.

triggerDetails_source :: Lens' TriggerDetails Text

Identifies the service that sourced the event. All events sourced from within AWS begin with "aws." Customer-generated events can have any value here, as long as it doesn't begin with "aws." We recommend the use of Java package-name style reverse domain-name strings.

triggerDetails_timestamp :: Lens' TriggerDetails UTCTime

The time that the incident was detected.

UpdateReplicationSetAction

data UpdateReplicationSetAction

Details used when updating the replication set.

See: newUpdateReplicationSetAction smart constructor.

Constructors

UpdateReplicationSetAction'

Fields

addRegionAction :: Maybe AddRegionAction

Details about the Region that you're adding to the replication set.

deleteRegionAction :: Maybe DeleteRegionAction

Details about the Region that you're deleting to the replication set.

newUpdateReplicationSetAction :: UpdateReplicationSetAction

Create a value of UpdateReplicationSetAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:addRegionAction:UpdateReplicationSetAction', updateReplicationSetAction_addRegionAction - Details about the Region that you're adding to the replication set.

$sel:deleteRegionAction:UpdateReplicationSetAction', updateReplicationSetAction_deleteRegionAction - Details about the Region that you're deleting to the replication set.

updateReplicationSetAction_addRegionAction :: Lens' UpdateReplicationSetAction (Maybe AddRegionAction)

Details about the Region that you're adding to the replication set.

updateReplicationSetAction_deleteRegionAction :: Lens' UpdateReplicationSetAction (Maybe DeleteRegionAction)

Details about the Region that you're deleting to the replication set.