Network.AWS.SSMIncidents.Types.ChatChannel

Description

Documentation

data ChatChannel

The AWS Chatbot chat channel used for collaboration during an incident.

See: newChatChannel smart constructor.

Constructors

ChatChannel'

Fields

empty :: Maybe EmptyChatChannel

Used to remove the chat channel from an incident record or response plan.

chatbotSns :: Maybe (NonEmpty Text)

The SNS targets that AWS Chatbot uses to notify the chat channel of updates to an incident. You can also make updates to the incident through the chat channel by using the SNS topics.

newChatChannel :: ChatChannel

Create a value of ChatChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:empty:ChatChannel', chatChannel_empty - Used to remove the chat channel from an incident record or response plan.

$sel:chatbotSns:ChatChannel', chatChannel_chatbotSns - The SNS targets that AWS Chatbot uses to notify the chat channel of updates to an incident. You can also make updates to the incident through the chat channel by using the SNS topics.

chatChannel_empty :: Lens' ChatChannel (Maybe EmptyChatChannel)

Used to remove the chat channel from an incident record or response plan.

chatChannel_chatbotSns :: Lens' ChatChannel (Maybe (NonEmpty Text))

The SNS targets that AWS Chatbot uses to notify the chat channel of updates to an incident. You can also make updates to the incident through the chat channel by using the SNS topics.