Network.AWS.SSMIncidents.Types.Condition

Description

Documentation

data Condition

A conditional statement with which to compare a value, after a timestamp, before a timestamp, or equal to a string or integer. If multiple conditions are specified, the conditionals become an ANDed statement. If multiple values are specified for a conditional, the values are ORd.

See: newCondition smart constructor.

Constructors

Condition'

Fields

after :: Maybe POSIX

After the specified timestamp.

equals :: Maybe AttributeValueList

The value is equal to the provided string or integer.

before :: Maybe POSIX

Before the specified timestamp

newCondition :: Condition

Create a value of Condition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:after:Condition', condition_after - After the specified timestamp.

$sel:equals:Condition', condition_equals - The value is equal to the provided string or integer.

$sel:before:Condition', condition_before - Before the specified timestamp

condition_after :: Lens' Condition (Maybe UTCTime)

After the specified timestamp.

condition_equals :: Lens' Condition (Maybe AttributeValueList)

The value is equal to the provided string or integer.

condition_before :: Lens' Condition (Maybe UTCTime)

Before the specified timestamp