Network.AWS.SSMIncidents.Types.IncidentRecordSummary

Description

Documentation

data IncidentRecordSummary

Details describing an incident record.

See: newIncidentRecordSummary smart constructor.

Constructors

IncidentRecordSummary'

Fields

resolvedTime :: Maybe POSIX

The time the incident was resolved.

arn :: Text

The Amazon Resource Name (ARN) of the incident.

creationTime :: POSIX

The time the incident was created.

impact :: Natural

Defines the impact to customers and applications.

incidentRecordSource :: IncidentRecordSource

What caused Incident Manager to create the incident.

status :: IncidentRecordStatus

The current status of the incident.

title :: Text

The title of the incident. This value is either provided by the response plan or overwritten on creation.

newIncidentRecordSummary

Arguments

:: Text
$sel:arn:IncidentRecordSummary'

-> UTCTime
$sel:creationTime:IncidentRecordSummary'

-> Natural
$sel:impact:IncidentRecordSummary'

-> IncidentRecordSource
$sel:incidentRecordSource:IncidentRecordSummary'

-> IncidentRecordStatus
$sel:status:IncidentRecordSummary'

-> Text
$sel:title:IncidentRecordSummary'

-> IncidentRecordSummary

Create a value of IncidentRecordSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolvedTime:IncidentRecordSummary', incidentRecordSummary_resolvedTime - The time the incident was resolved.

$sel:arn:IncidentRecordSummary', incidentRecordSummary_arn - The Amazon Resource Name (ARN) of the incident.

$sel:creationTime:IncidentRecordSummary', incidentRecordSummary_creationTime - The time the incident was created.

$sel:impact:IncidentRecordSummary', incidentRecordSummary_impact - Defines the impact to customers and applications.

$sel:incidentRecordSource:IncidentRecordSummary', incidentRecordSummary_incidentRecordSource - What caused Incident Manager to create the incident.

$sel:status:IncidentRecordSummary', incidentRecordSummary_status - The current status of the incident.

$sel:title:IncidentRecordSummary', incidentRecordSummary_title - The title of the incident. This value is either provided by the response plan or overwritten on creation.

incidentRecordSummary_resolvedTime :: Lens' IncidentRecordSummary (Maybe UTCTime)

The time the incident was resolved.

incidentRecordSummary_arn :: Lens' IncidentRecordSummary Text

The Amazon Resource Name (ARN) of the incident.

incidentRecordSummary_creationTime :: Lens' IncidentRecordSummary UTCTime

The time the incident was created.

incidentRecordSummary_impact :: Lens' IncidentRecordSummary Natural

Defines the impact to customers and applications.

incidentRecordSummary_incidentRecordSource :: Lens' IncidentRecordSummary IncidentRecordSource

What caused Incident Manager to create the incident.

incidentRecordSummary_status :: Lens' IncidentRecordSummary IncidentRecordStatus

The current status of the incident.

incidentRecordSummary_title :: Lens' IncidentRecordSummary Text

The title of the incident. This value is either provided by the response plan or overwritten on creation.