Network.AWS.SSMIncidents.Types.ResourcePolicy

Description

Documentation

data ResourcePolicy

The resource policy that allows Incident Manager to perform actions on resources on your behalf.

See: newResourcePolicy smart constructor.

Constructors

ResourcePolicy'

Fields

policyDocument :: Text

The JSON blob that describes the policy.

policyId :: Text

The ID of the resource policy.

ramResourceShareRegion :: Text

The Region that policy allows resources to be used in.

newResourcePolicy

Arguments

:: Text
$sel:policyDocument:ResourcePolicy'

-> Text
$sel:policyId:ResourcePolicy'

-> Text
$sel:ramResourceShareRegion:ResourcePolicy'

-> ResourcePolicy

Create a value of ResourcePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyDocument:ResourcePolicy', resourcePolicy_policyDocument - The JSON blob that describes the policy.

$sel:policyId:ResourcePolicy', resourcePolicy_policyId - The ID of the resource policy.

$sel:ramResourceShareRegion:ResourcePolicy', resourcePolicy_ramResourceShareRegion - The Region that policy allows resources to be used in.

resourcePolicy_policyDocument :: Lens' ResourcePolicy Text

The JSON blob that describes the policy.

resourcePolicy_policyId :: Lens' ResourcePolicy Text

The ID of the resource policy.

resourcePolicy_ramResourceShareRegion :: Lens' ResourcePolicy Text

The Region that policy allows resources to be used in.