Network.AWS.SSMIncidents.UpdateResponsePlan

Description

Updates the specified response plan.

Creating a Request

data UpdateResponsePlan

See: newUpdateResponsePlan smart constructor.

Constructors

UpdateResponsePlan'

Fields

incidentTemplateImpact :: Maybe Natural

Defines the impact to the customers. Providing an impact overwrites the impact provided by a response plan.

Possible impacts:

5 - Severe impact

4 - High impact

3 - Medium impact

2 - Low impact

1 - No impact

clientToken :: Maybe Text

A token ensuring that the action is called only once with the specified details.

actions :: Maybe [Action]

The actions that this response plan takes at the beginning of an incident.

incidentTemplateSummary :: Maybe Text

A brief summary of the incident. This typically contains what has happened, what's currently happening, and next steps.

displayName :: Maybe Text

The long format name of the response plan. Can't contain spaces.

chatChannel :: Maybe ChatChannel

The AWS Chatbot chat channel used for collaboration during an incident.

Use the empty structure to remove the chat channel from the response plan.

incidentTemplateDedupeString :: Maybe Text

Used to create only one incident record for an incident.

incidentTemplateTitle :: Maybe Text

The short format name of the incident. Can't contain spaces.

engagements :: Maybe [Text]

The contacts and escalation plans that Incident Manager engages at the start of the incident.

incidentTemplateNotificationTargets :: Maybe [NotificationTargetItem]

The SNS targets that are notified when updates are made to an incident.

arn :: Text

The Amazon Resource Name (ARN) of the response plan.

newUpdateResponsePlan

Arguments

:: Text
$sel:arn:UpdateResponsePlan'

-> UpdateResponsePlan

Create a value of UpdateResponsePlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:incidentTemplateImpact:UpdateResponsePlan', updateResponsePlan_incidentTemplateImpact - Defines the impact to the customers. Providing an impact overwrites the impact provided by a response plan.

Possible impacts:

5 - Severe impact

4 - High impact

3 - Medium impact

2 - Low impact

1 - No impact

$sel:clientToken:UpdateResponsePlan', updateResponsePlan_clientToken - A token ensuring that the action is called only once with the specified details.

$sel:actions:UpdateResponsePlan', updateResponsePlan_actions - The actions that this response plan takes at the beginning of an incident.

$sel:incidentTemplateSummary:UpdateResponsePlan', updateResponsePlan_incidentTemplateSummary - A brief summary of the incident. This typically contains what has happened, what's currently happening, and next steps.

$sel:displayName:UpdateResponsePlan', updateResponsePlan_displayName - The long format name of the response plan. Can't contain spaces.

$sel:chatChannel:UpdateResponsePlan', updateResponsePlan_chatChannel - The AWS Chatbot chat channel used for collaboration during an incident.

Use the empty structure to remove the chat channel from the response plan.

$sel:incidentTemplateDedupeString:UpdateResponsePlan', updateResponsePlan_incidentTemplateDedupeString - Used to create only one incident record for an incident.

$sel:incidentTemplateTitle:UpdateResponsePlan', updateResponsePlan_incidentTemplateTitle - The short format name of the incident. Can't contain spaces.

$sel:engagements:UpdateResponsePlan', updateResponsePlan_engagements - The contacts and escalation plans that Incident Manager engages at the start of the incident.

$sel:incidentTemplateNotificationTargets:UpdateResponsePlan', updateResponsePlan_incidentTemplateNotificationTargets - The SNS targets that are notified when updates are made to an incident.

$sel:arn:UpdateResponsePlan', updateResponsePlan_arn - The Amazon Resource Name (ARN) of the response plan.

Request Lenses

updateResponsePlan_incidentTemplateImpact :: Lens' UpdateResponsePlan (Maybe Natural)

Defines the impact to the customers. Providing an impact overwrites the impact provided by a response plan.

Possible impacts:

5 - Severe impact

4 - High impact

3 - Medium impact

2 - Low impact

1 - No impact

updateResponsePlan_clientToken :: Lens' UpdateResponsePlan (Maybe Text)

A token ensuring that the action is called only once with the specified details.

updateResponsePlan_actions :: Lens' UpdateResponsePlan (Maybe [Action])

The actions that this response plan takes at the beginning of an incident.

updateResponsePlan_incidentTemplateSummary :: Lens' UpdateResponsePlan (Maybe Text)

A brief summary of the incident. This typically contains what has happened, what's currently happening, and next steps.

updateResponsePlan_displayName :: Lens' UpdateResponsePlan (Maybe Text)

The long format name of the response plan. Can't contain spaces.

updateResponsePlan_chatChannel :: Lens' UpdateResponsePlan (Maybe ChatChannel)

The AWS Chatbot chat channel used for collaboration during an incident.

Use the empty structure to remove the chat channel from the response plan.

updateResponsePlan_incidentTemplateDedupeString :: Lens' UpdateResponsePlan (Maybe Text)

Used to create only one incident record for an incident.

updateResponsePlan_incidentTemplateTitle :: Lens' UpdateResponsePlan (Maybe Text)

The short format name of the incident. Can't contain spaces.

updateResponsePlan_engagements :: Lens' UpdateResponsePlan (Maybe [Text])

The contacts and escalation plans that Incident Manager engages at the start of the incident.

updateResponsePlan_incidentTemplateNotificationTargets :: Lens' UpdateResponsePlan (Maybe [NotificationTargetItem])

The SNS targets that are notified when updates are made to an incident.

updateResponsePlan_arn :: Lens' UpdateResponsePlan Text

The Amazon Resource Name (ARN) of the response plan.

Destructuring the Response

data UpdateResponsePlanResponse

See: newUpdateResponsePlanResponse smart constructor.

Constructors

UpdateResponsePlanResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateResponsePlanResponse

Arguments

:: Int
$sel:httpStatus:UpdateResponsePlanResponse'

-> UpdateResponsePlanResponse

Create a value of UpdateResponsePlanResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateResponsePlanResponse', updateResponsePlanResponse_httpStatus - The response's http status code.

Response Lenses

updateResponsePlanResponse_httpStatus :: Lens' UpdateResponsePlanResponse Int

The response's http status code.