Network.AWS.SSO.Lens

Description

Operations

Logout

logout_accessToken :: Lens' Logout Text

The token issued by the CreateToken API call. For more information, see CreateToken in the AWS SSO OIDC API Reference Guide.

GetRoleCredentials

getRoleCredentials_roleName :: Lens' GetRoleCredentials Text

The friendly name of the role that is assigned to the user.

getRoleCredentials_accountId :: Lens' GetRoleCredentials Text

The identifier for the AWS account that is assigned to the user.

getRoleCredentials_accessToken :: Lens' GetRoleCredentials Text

The token issued by the CreateToken API call. For more information, see CreateToken in the AWS SSO OIDC API Reference Guide.

getRoleCredentialsResponse_roleCredentials :: Lens' GetRoleCredentialsResponse (Maybe RoleCredentials)

The credentials for the role that is assigned to the user.

getRoleCredentialsResponse_httpStatus :: Lens' GetRoleCredentialsResponse Int

The response's http status code.

ListAccounts

listAccounts_nextToken :: Lens' ListAccounts (Maybe Text)

(Optional) When requesting subsequent pages, this is the page token from the previous response output.

listAccounts_maxResults :: Lens' ListAccounts (Maybe Natural)

This is the number of items clients can request per page.

listAccounts_accessToken :: Lens' ListAccounts Text

The token issued by the CreateToken API call. For more information, see CreateToken in the AWS SSO OIDC API Reference Guide.

listAccountsResponse_accountList :: Lens' ListAccountsResponse (Maybe [AccountInfo])

A paginated response with the list of account information and the next token if more results are available.

listAccountsResponse_nextToken :: Lens' ListAccountsResponse (Maybe Text)

The page token client that is used to retrieve the list of accounts.

listAccountsResponse_httpStatus :: Lens' ListAccountsResponse Int

The response's http status code.

ListAccountRoles

listAccountRoles_nextToken :: Lens' ListAccountRoles (Maybe Text)

The page token from the previous response output when you request subsequent pages.

listAccountRoles_maxResults :: Lens' ListAccountRoles (Maybe Natural)

The number of items that clients can request per page.

listAccountRoles_accessToken :: Lens' ListAccountRoles Text

The token issued by the CreateToken API call. For more information, see CreateToken in the AWS SSO OIDC API Reference Guide.

listAccountRoles_accountId :: Lens' ListAccountRoles Text

The identifier for the AWS account that is assigned to the user.

listAccountRolesResponse_roleList :: Lens' ListAccountRolesResponse (Maybe [RoleInfo])

A paginated response with the list of roles and the next token if more results are available.

listAccountRolesResponse_nextToken :: Lens' ListAccountRolesResponse (Maybe Text)

The page token client that is used to retrieve the list of accounts.

listAccountRolesResponse_httpStatus :: Lens' ListAccountRolesResponse Int

The response's http status code.

Types

AccountInfo

accountInfo_accountName :: Lens' AccountInfo (Maybe Text)

The display name of the AWS account that is assigned to the user.

accountInfo_accountId :: Lens' AccountInfo (Maybe Text)

The identifier of the AWS account that is assigned to the user.

accountInfo_emailAddress :: Lens' AccountInfo (Maybe Text)

The email address of the AWS account that is assigned to the user.

RoleCredentials

roleCredentials_secretAccessKey :: Lens' RoleCredentials (Maybe Text)

The key that is used to sign the request. For more information, see Using Temporary Security Credentials to Request Access to AWS Resources in the AWS IAM User Guide.

roleCredentials_sessionToken :: Lens' RoleCredentials (Maybe Text)

The token used for temporary credentials. For more information, see Using Temporary Security Credentials to Request Access to AWS Resources in the AWS IAM User Guide.

roleCredentials_expiration :: Lens' RoleCredentials (Maybe Integer)

The date on which temporary security credentials expire.

roleCredentials_accessKeyId :: Lens' RoleCredentials (Maybe Text)

The identifier used for the temporary security credentials. For more information, see Using Temporary Security Credentials to Request Access to AWS Resources in the AWS IAM User Guide.

RoleInfo

roleInfo_roleName :: Lens' RoleInfo (Maybe Text)

The friendly name of the role that is assigned to the user.

roleInfo_accountId :: Lens' RoleInfo (Maybe Text)

The identifier of the AWS account assigned to the user.