Network.AWS.SSO.ListAccountRoles

Description

Lists all roles that are assigned to the user for a given AWS account.

This operation returns paginated results.

Creating a Request

data ListAccountRoles

See: newListAccountRoles smart constructor.

Constructors

ListAccountRoles'

Fields

nextToken :: Maybe Text

The page token from the previous response output when you request subsequent pages.

maxResults :: Maybe Natural

The number of items that clients can request per page.

accessToken :: Sensitive Text

The token issued by the CreateToken API call. For more information, see CreateToken in the AWS SSO OIDC API Reference Guide.

accountId :: Text

The identifier for the AWS account that is assigned to the user.

newListAccountRoles

Arguments

:: Text
$sel:accessToken:ListAccountRoles'

-> Text
$sel:accountId:ListAccountRoles'

-> ListAccountRoles

Create a value of ListAccountRoles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAccountRoles', listAccountRoles_nextToken - The page token from the previous response output when you request subsequent pages.

$sel:maxResults:ListAccountRoles', listAccountRoles_maxResults - The number of items that clients can request per page.

$sel:accessToken:ListAccountRoles', listAccountRoles_accessToken - The token issued by the CreateToken API call. For more information, see CreateToken in the AWS SSO OIDC API Reference Guide.

$sel:accountId:ListAccountRoles', listAccountRoles_accountId - The identifier for the AWS account that is assigned to the user.

Request Lenses

listAccountRoles_nextToken :: Lens' ListAccountRoles (Maybe Text)

The page token from the previous response output when you request subsequent pages.

listAccountRoles_maxResults :: Lens' ListAccountRoles (Maybe Natural)

The number of items that clients can request per page.

listAccountRoles_accessToken :: Lens' ListAccountRoles Text

The token issued by the CreateToken API call. For more information, see CreateToken in the AWS SSO OIDC API Reference Guide.

listAccountRoles_accountId :: Lens' ListAccountRoles Text

The identifier for the AWS account that is assigned to the user.

Destructuring the Response

data ListAccountRolesResponse

See: newListAccountRolesResponse smart constructor.

Constructors

ListAccountRolesResponse'

Fields

roleList :: Maybe [RoleInfo]

A paginated response with the list of roles and the next token if more results are available.

nextToken :: Maybe Text

The page token client that is used to retrieve the list of accounts.

httpStatus :: Int

The response's http status code.

newListAccountRolesResponse

Arguments

:: Int
$sel:httpStatus:ListAccountRolesResponse'

-> ListAccountRolesResponse

Create a value of ListAccountRolesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleList:ListAccountRolesResponse', listAccountRolesResponse_roleList - A paginated response with the list of roles and the next token if more results are available.

$sel:nextToken:ListAccountRoles', listAccountRolesResponse_nextToken - The page token client that is used to retrieve the list of accounts.

$sel:httpStatus:ListAccountRolesResponse', listAccountRolesResponse_httpStatus - The response's http status code.

Response Lenses

listAccountRolesResponse_roleList :: Lens' ListAccountRolesResponse (Maybe [RoleInfo])

A paginated response with the list of roles and the next token if more results are available.

listAccountRolesResponse_nextToken :: Lens' ListAccountRolesResponse (Maybe Text)

The page token client that is used to retrieve the list of accounts.

listAccountRolesResponse_httpStatus :: Lens' ListAccountRolesResponse Int

The response's http status code.