Network.AWS.SSOAdmin.CreatePermissionSet

Description

Creates a permission set within a specified SSO instance.

To grant users and groups access to Amazon Web Services account resources, use CreateAccountAssignment .

Creating a Request

data CreatePermissionSet

See: newCreatePermissionSet smart constructor.

Constructors

CreatePermissionSet'

Fields

relayState :: Maybe Text

Used to redirect users within the application during the federation authentication process.

sessionDuration :: Maybe Text

The length of time that the application user sessions are valid in the ISO-8601 standard.

description :: Maybe Text

The description of the PermissionSet.

tags :: Maybe [Tag]

The tags to attach to the new PermissionSet.

name :: Text

The name of the PermissionSet.

instanceArn :: Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

newCreatePermissionSet

Arguments

:: Text
$sel:name:CreatePermissionSet'

-> Text
$sel:instanceArn:CreatePermissionSet'

-> CreatePermissionSet

Create a value of CreatePermissionSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:relayState:CreatePermissionSet', createPermissionSet_relayState - Used to redirect users within the application during the federation authentication process.

$sel:sessionDuration:CreatePermissionSet', createPermissionSet_sessionDuration - The length of time that the application user sessions are valid in the ISO-8601 standard.

$sel:description:CreatePermissionSet', createPermissionSet_description - The description of the PermissionSet.

$sel:tags:CreatePermissionSet', createPermissionSet_tags - The tags to attach to the new PermissionSet.

$sel:name:CreatePermissionSet', createPermissionSet_name - The name of the PermissionSet.

$sel:instanceArn:CreatePermissionSet', createPermissionSet_instanceArn - The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

Request Lenses

createPermissionSet_relayState :: Lens' CreatePermissionSet (Maybe Text)

Used to redirect users within the application during the federation authentication process.

createPermissionSet_sessionDuration :: Lens' CreatePermissionSet (Maybe Text)

The length of time that the application user sessions are valid in the ISO-8601 standard.

createPermissionSet_description :: Lens' CreatePermissionSet (Maybe Text)

The description of the PermissionSet.

createPermissionSet_tags :: Lens' CreatePermissionSet (Maybe [Tag])

The tags to attach to the new PermissionSet.

createPermissionSet_name :: Lens' CreatePermissionSet Text

The name of the PermissionSet.

createPermissionSet_instanceArn :: Lens' CreatePermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

Destructuring the Response

data CreatePermissionSetResponse

See: newCreatePermissionSetResponse smart constructor.

Constructors

CreatePermissionSetResponse'

Fields

permissionSet :: Maybe PermissionSet

Defines the level of access on an Amazon Web Services account.

httpStatus :: Int

The response's http status code.

newCreatePermissionSetResponse

Arguments

:: Int
$sel:httpStatus:CreatePermissionSetResponse'

-> CreatePermissionSetResponse

Create a value of CreatePermissionSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:permissionSet:CreatePermissionSetResponse', createPermissionSetResponse_permissionSet - Defines the level of access on an Amazon Web Services account.

$sel:httpStatus:CreatePermissionSetResponse', createPermissionSetResponse_httpStatus - The response's http status code.

Response Lenses

createPermissionSetResponse_permissionSet :: Lens' CreatePermissionSetResponse (Maybe PermissionSet)

Defines the level of access on an Amazon Web Services account.

createPermissionSetResponse_httpStatus :: Lens' CreatePermissionSetResponse Int

The response's http status code.