Network.AWS.SSOAdmin.DeletePermissionSet

Description

Deletes the specified permission set.

Creating a Request

data DeletePermissionSet

See: newDeletePermissionSet smart constructor.

Constructors

DeletePermissionSet'

Fields

instanceArn :: Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

permissionSetArn :: Text

The ARN of the permission set that should be deleted.

newDeletePermissionSet

Arguments

:: Text
$sel:instanceArn:DeletePermissionSet'

-> Text
$sel:permissionSetArn:DeletePermissionSet'

-> DeletePermissionSet

Create a value of DeletePermissionSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceArn:DeletePermissionSet', deletePermissionSet_instanceArn - The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

$sel:permissionSetArn:DeletePermissionSet', deletePermissionSet_permissionSetArn - The ARN of the permission set that should be deleted.

Request Lenses

deletePermissionSet_instanceArn :: Lens' DeletePermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

deletePermissionSet_permissionSetArn :: Lens' DeletePermissionSet Text

The ARN of the permission set that should be deleted.

Destructuring the Response

data DeletePermissionSetResponse

See: newDeletePermissionSetResponse smart constructor.

Constructors

DeletePermissionSetResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeletePermissionSetResponse

Arguments

:: Int
$sel:httpStatus:DeletePermissionSetResponse'

-> DeletePermissionSetResponse

Create a value of DeletePermissionSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeletePermissionSetResponse', deletePermissionSetResponse_httpStatus - The response's http status code.

Response Lenses

deletePermissionSetResponse_httpStatus :: Lens' DeletePermissionSetResponse Int

The response's http status code.