Network.AWS.SSOAdmin.DescribeAccountAssignmentDeletionStatus

Description

Describes the status of the assignment deletion request.

Creating a Request

data DescribeAccountAssignmentDeletionStatus

See: newDescribeAccountAssignmentDeletionStatus smart constructor.

Constructors

DescribeAccountAssignmentDeletionStatus'

Fields

instanceArn :: Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

accountAssignmentDeletionRequestId :: Text

The identifier that is used to track the request operation progress.

newDescribeAccountAssignmentDeletionStatus

Arguments

:: Text
$sel:instanceArn:DescribeAccountAssignmentDeletionStatus'

-> Text
$sel:accountAssignmentDeletionRequestId:DescribeAccountAssignmentDeletionStatus'

-> DescribeAccountAssignmentDeletionStatus

Create a value of DescribeAccountAssignmentDeletionStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceArn:DescribeAccountAssignmentDeletionStatus', describeAccountAssignmentDeletionStatus_instanceArn - The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

$sel:accountAssignmentDeletionRequestId:DescribeAccountAssignmentDeletionStatus', describeAccountAssignmentDeletionStatus_accountAssignmentDeletionRequestId - The identifier that is used to track the request operation progress.

Request Lenses

describeAccountAssignmentDeletionStatus_instanceArn :: Lens' DescribeAccountAssignmentDeletionStatus Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

describeAccountAssignmentDeletionStatus_accountAssignmentDeletionRequestId :: Lens' DescribeAccountAssignmentDeletionStatus Text

The identifier that is used to track the request operation progress.

Destructuring the Response

data DescribeAccountAssignmentDeletionStatusResponse

See: newDescribeAccountAssignmentDeletionStatusResponse smart constructor.

Constructors

DescribeAccountAssignmentDeletionStatusResponse'

Fields

accountAssignmentDeletionStatus :: Maybe AccountAssignmentOperationStatus

The status object for the account assignment deletion operation.

httpStatus :: Int

The response's http status code.

newDescribeAccountAssignmentDeletionStatusResponse

Arguments

:: Int
$sel:httpStatus:DescribeAccountAssignmentDeletionStatusResponse'

-> DescribeAccountAssignmentDeletionStatusResponse

Create a value of DescribeAccountAssignmentDeletionStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountAssignmentDeletionStatus:DescribeAccountAssignmentDeletionStatusResponse', describeAccountAssignmentDeletionStatusResponse_accountAssignmentDeletionStatus - The status object for the account assignment deletion operation.

$sel:httpStatus:DescribeAccountAssignmentDeletionStatusResponse', describeAccountAssignmentDeletionStatusResponse_httpStatus - The response's http status code.

Response Lenses

describeAccountAssignmentDeletionStatusResponse_accountAssignmentDeletionStatus :: Lens' DescribeAccountAssignmentDeletionStatusResponse (Maybe AccountAssignmentOperationStatus)

The status object for the account assignment deletion operation.

describeAccountAssignmentDeletionStatusResponse_httpStatus :: Lens' DescribeAccountAssignmentDeletionStatusResponse Int

The response's http status code.