Network.AWS.SSOAdmin.DescribeInstanceAccessControlAttributeConfiguration

Description

Returns the list of Amazon Web Services SSO identity store attributes that have been configured to work with attributes-based access control (ABAC) for the specified Amazon Web Services SSO instance. This will not return attributes configured and sent by an external identity provider. For more information about ABAC, see Attribute-Based Access Control in the Amazon Web Services SSO User Guide.

Creating a Request

data DescribeInstanceAccessControlAttributeConfiguration

See: newDescribeInstanceAccessControlAttributeConfiguration smart constructor.

Constructors

DescribeInstanceAccessControlAttributeConfiguration'

Fields

instanceArn :: Text

The ARN of the SSO instance under which the operation will be executed.

newDescribeInstanceAccessControlAttributeConfiguration

Arguments

:: Text
$sel:instanceArn:DescribeInstanceAccessControlAttributeConfiguration'

-> DescribeInstanceAccessControlAttributeConfiguration

Create a value of DescribeInstanceAccessControlAttributeConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceArn:DescribeInstanceAccessControlAttributeConfiguration', describeInstanceAccessControlAttributeConfiguration_instanceArn - The ARN of the SSO instance under which the operation will be executed.

Request Lenses

describeInstanceAccessControlAttributeConfiguration_instanceArn :: Lens' DescribeInstanceAccessControlAttributeConfiguration Text

The ARN of the SSO instance under which the operation will be executed.

Destructuring the Response

data DescribeInstanceAccessControlAttributeConfigurationResponse

See: newDescribeInstanceAccessControlAttributeConfigurationResponse smart constructor.

Constructors

DescribeInstanceAccessControlAttributeConfigurationResponse'

Fields

status :: Maybe InstanceAccessControlAttributeConfigurationStatus

The status of the attribute configuration process.

instanceAccessControlAttributeConfiguration :: Maybe InstanceAccessControlAttributeConfiguration

Gets the list of Amazon Web Services SSO identity store attributes that have been added to your ABAC configuration.

statusReason :: Maybe Text

Provides more details about the current status of the specified attribute.

httpStatus :: Int

The response's http status code.

newDescribeInstanceAccessControlAttributeConfigurationResponse

Arguments

:: Int
$sel:httpStatus:DescribeInstanceAccessControlAttributeConfigurationResponse'

-> DescribeInstanceAccessControlAttributeConfigurationResponse

Create a value of DescribeInstanceAccessControlAttributeConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeInstanceAccessControlAttributeConfigurationResponse', describeInstanceAccessControlAttributeConfigurationResponse_status - The status of the attribute configuration process.

$sel:instanceAccessControlAttributeConfiguration:DescribeInstanceAccessControlAttributeConfigurationResponse', describeInstanceAccessControlAttributeConfigurationResponse_instanceAccessControlAttributeConfiguration - Gets the list of Amazon Web Services SSO identity store attributes that have been added to your ABAC configuration.

$sel:statusReason:DescribeInstanceAccessControlAttributeConfigurationResponse', describeInstanceAccessControlAttributeConfigurationResponse_statusReason - Provides more details about the current status of the specified attribute.

$sel:httpStatus:DescribeInstanceAccessControlAttributeConfigurationResponse', describeInstanceAccessControlAttributeConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

describeInstanceAccessControlAttributeConfigurationResponse_status :: Lens' DescribeInstanceAccessControlAttributeConfigurationResponse (Maybe InstanceAccessControlAttributeConfigurationStatus)

The status of the attribute configuration process.

describeInstanceAccessControlAttributeConfigurationResponse_instanceAccessControlAttributeConfiguration :: Lens' DescribeInstanceAccessControlAttributeConfigurationResponse (Maybe InstanceAccessControlAttributeConfiguration)

Gets the list of Amazon Web Services SSO identity store attributes that have been added to your ABAC configuration.

describeInstanceAccessControlAttributeConfigurationResponse_statusReason :: Lens' DescribeInstanceAccessControlAttributeConfigurationResponse (Maybe Text)

Provides more details about the current status of the specified attribute.

describeInstanceAccessControlAttributeConfigurationResponse_httpStatus :: Lens' DescribeInstanceAccessControlAttributeConfigurationResponse Int

The response's http status code.