Network.AWS.SSOAdmin.ListPermissionSets

Description

Lists the PermissionSets in an SSO instance.

This operation returns paginated results.

Creating a Request

data ListPermissionSets

See: newListPermissionSets smart constructor.

Constructors

ListPermissionSets'

Fields

nextToken :: Maybe Text

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

maxResults :: Maybe Natural

The maximum number of results to display for the assignment.

instanceArn :: Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

newListPermissionSets

Arguments

:: Text
$sel:instanceArn:ListPermissionSets'

-> ListPermissionSets

Create a value of ListPermissionSets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPermissionSets', listPermissionSets_nextToken - The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

$sel:maxResults:ListPermissionSets', listPermissionSets_maxResults - The maximum number of results to display for the assignment.

$sel:instanceArn:ListPermissionSets', listPermissionSets_instanceArn - The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

Request Lenses

listPermissionSets_nextToken :: Lens' ListPermissionSets (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listPermissionSets_maxResults :: Lens' ListPermissionSets (Maybe Natural)

The maximum number of results to display for the assignment.

listPermissionSets_instanceArn :: Lens' ListPermissionSets Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

Destructuring the Response

data ListPermissionSetsResponse

See: newListPermissionSetsResponse smart constructor.

Constructors

ListPermissionSetsResponse'

Fields

permissionSets :: Maybe [Text]

Defines the level of access on an Amazon Web Services account.

nextToken :: Maybe Text

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

httpStatus :: Int

The response's http status code.

newListPermissionSetsResponse

Arguments

:: Int
$sel:httpStatus:ListPermissionSetsResponse'

-> ListPermissionSetsResponse

Create a value of ListPermissionSetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:permissionSets:ListPermissionSetsResponse', listPermissionSetsResponse_permissionSets - Defines the level of access on an Amazon Web Services account.

$sel:nextToken:ListPermissionSets', listPermissionSetsResponse_nextToken - The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

$sel:httpStatus:ListPermissionSetsResponse', listPermissionSetsResponse_httpStatus - The response's http status code.

Response Lenses

listPermissionSetsResponse_permissionSets :: Lens' ListPermissionSetsResponse (Maybe [Text])

Defines the level of access on an Amazon Web Services account.

listPermissionSetsResponse_nextToken :: Lens' ListPermissionSetsResponse (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listPermissionSetsResponse_httpStatus :: Lens' ListPermissionSetsResponse Int

The response's http status code.