Network.AWS.SSOAdmin.ListPermissionSetsProvisionedToAccount

Description

Lists all the permission sets that are provisioned to a specified Amazon Web Services account.

This operation returns paginated results.

Creating a Request

data ListPermissionSetsProvisionedToAccount

See: newListPermissionSetsProvisionedToAccount smart constructor.

Constructors

ListPermissionSetsProvisionedToAccount'

Fields

provisioningStatus :: Maybe ProvisioningStatus

The status object for the permission set provisioning operation.

nextToken :: Maybe Text

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

maxResults :: Maybe Natural

The maximum number of results to display for the assignment.

instanceArn :: Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

accountId :: Text

The identifier of the Amazon Web Services account from which to list the assignments.

newListPermissionSetsProvisionedToAccount

Arguments

:: Text
$sel:instanceArn:ListPermissionSetsProvisionedToAccount'

-> Text
$sel:accountId:ListPermissionSetsProvisionedToAccount'

-> ListPermissionSetsProvisionedToAccount

Create a value of ListPermissionSetsProvisionedToAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:provisioningStatus:ListPermissionSetsProvisionedToAccount', listPermissionSetsProvisionedToAccount_provisioningStatus - The status object for the permission set provisioning operation.

$sel:nextToken:ListPermissionSetsProvisionedToAccount', listPermissionSetsProvisionedToAccount_nextToken - The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

$sel:maxResults:ListPermissionSetsProvisionedToAccount', listPermissionSetsProvisionedToAccount_maxResults - The maximum number of results to display for the assignment.

$sel:instanceArn:ListPermissionSetsProvisionedToAccount', listPermissionSetsProvisionedToAccount_instanceArn - The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

$sel:accountId:ListPermissionSetsProvisionedToAccount', listPermissionSetsProvisionedToAccount_accountId - The identifier of the Amazon Web Services account from which to list the assignments.

Request Lenses

listPermissionSetsProvisionedToAccount_provisioningStatus :: Lens' ListPermissionSetsProvisionedToAccount (Maybe ProvisioningStatus)

The status object for the permission set provisioning operation.

listPermissionSetsProvisionedToAccount_nextToken :: Lens' ListPermissionSetsProvisionedToAccount (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listPermissionSetsProvisionedToAccount_maxResults :: Lens' ListPermissionSetsProvisionedToAccount (Maybe Natural)

The maximum number of results to display for the assignment.

listPermissionSetsProvisionedToAccount_instanceArn :: Lens' ListPermissionSetsProvisionedToAccount Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

listPermissionSetsProvisionedToAccount_accountId :: Lens' ListPermissionSetsProvisionedToAccount Text

The identifier of the Amazon Web Services account from which to list the assignments.

Destructuring the Response

data ListPermissionSetsProvisionedToAccountResponse

See: newListPermissionSetsProvisionedToAccountResponse smart constructor.

Constructors

ListPermissionSetsProvisionedToAccountResponse'

Fields

permissionSets :: Maybe [Text]

Defines the level of access that an Amazon Web Services account has.

nextToken :: Maybe Text

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

httpStatus :: Int

The response's http status code.

newListPermissionSetsProvisionedToAccountResponse

Arguments

:: Int
$sel:httpStatus:ListPermissionSetsProvisionedToAccountResponse'

-> ListPermissionSetsProvisionedToAccountResponse

Create a value of ListPermissionSetsProvisionedToAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:permissionSets:ListPermissionSetsProvisionedToAccountResponse', listPermissionSetsProvisionedToAccountResponse_permissionSets - Defines the level of access that an Amazon Web Services account has.

$sel:nextToken:ListPermissionSetsProvisionedToAccount', listPermissionSetsProvisionedToAccountResponse_nextToken - The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

$sel:httpStatus:ListPermissionSetsProvisionedToAccountResponse', listPermissionSetsProvisionedToAccountResponse_httpStatus - The response's http status code.

Response Lenses

listPermissionSetsProvisionedToAccountResponse_permissionSets :: Lens' ListPermissionSetsProvisionedToAccountResponse (Maybe [Text])

Defines the level of access that an Amazon Web Services account has.

listPermissionSetsProvisionedToAccountResponse_nextToken :: Lens' ListPermissionSetsProvisionedToAccountResponse (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listPermissionSetsProvisionedToAccountResponse_httpStatus :: Lens' ListPermissionSetsProvisionedToAccountResponse Int

The response's http status code.