Network.AWS.SSOAdmin.Types

Description

Service Configuration

defaultService :: Service

API version 2020-07-20 of the Amazon Single Sign-On Admin SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The request failed because it contains a syntax error.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Occurs when a conflict with a previous successful write is detected. This generally occurs when the previous write did not have time to propagate to the host serving the current request. A retry (with appropriate backoff logic) is the recommended response to this exception.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the principal has crossed the permitted number of resources that can be created.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the principal has crossed the throttling limits of the API operations.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The request processing has failed because of an unknown error, exception, or failure with an internal server.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that a requested resource is not found.

InstanceAccessControlAttributeConfigurationStatus

newtype InstanceAccessControlAttributeConfigurationStatus

Constructors

InstanceAccessControlAttributeConfigurationStatus'

Fields

fromInstanceAccessControlAttributeConfigurationStatus :: Text

Bundled Patterns

pattern InstanceAccessControlAttributeConfigurationStatus_CREATION_FAILED :: InstanceAccessControlAttributeConfigurationStatus

pattern InstanceAccessControlAttributeConfigurationStatus_CREATION_IN_PROGRESS :: InstanceAccessControlAttributeConfigurationStatus

pattern InstanceAccessControlAttributeConfigurationStatus_ENABLED :: InstanceAccessControlAttributeConfigurationStatus

PrincipalType

newtype PrincipalType

Constructors

PrincipalType'

Fields

fromPrincipalType :: Text

Bundled Patterns

pattern PrincipalType_GROUP :: PrincipalType

pattern PrincipalType_USER :: PrincipalType

ProvisionTargetType

newtype ProvisionTargetType

Constructors

ProvisionTargetType'

Fields

fromProvisionTargetType :: Text

Bundled Patterns

pattern ProvisionTargetType_ALL_PROVISIONED_ACCOUNTS :: ProvisionTargetType

pattern ProvisionTargetType_AWS_ACCOUNT :: ProvisionTargetType

ProvisioningStatus

newtype ProvisioningStatus

Constructors

ProvisioningStatus'

Fields

fromProvisioningStatus :: Text

Bundled Patterns

pattern ProvisioningStatus_LATEST_PERMISSION_SET_NOT_PROVISIONED :: ProvisioningStatus

pattern ProvisioningStatus_LATEST_PERMISSION_SET_PROVISIONED :: ProvisioningStatus

StatusValues

newtype StatusValues

Constructors

StatusValues'

Fields

fromStatusValues :: Text

Bundled Patterns

pattern StatusValues_FAILED :: StatusValues

pattern StatusValues_IN_PROGRESS :: StatusValues

pattern StatusValues_SUCCEEDED :: StatusValues

TargetType

newtype TargetType

Constructors

TargetType'

Fields

fromTargetType :: Text

Bundled Patterns

pattern TargetType_AWS_ACCOUNT :: TargetType

AccessControlAttribute

data AccessControlAttribute

These are Amazon Web Services SSO identity store attributes that you can configure for use in attributes-based access control (ABAC). You can create permissions policies that determine who can access your Amazon Web Services resources based upon the configured attribute values. When you enable ABAC and specify AccessControlAttributes, Amazon Web Services SSO passes the attribute values of the authenticated user into IAM for use in policy evaluation.

See: newAccessControlAttribute smart constructor.

Constructors

AccessControlAttribute'

Fields

key :: Text

The name of the attribute associated with your identities in your identity source. This is used to map a specified attribute in your identity source with an attribute in Amazon Web Services SSO.

value :: AccessControlAttributeValue

The value used for mapping a specified attribute to an identity source.

newAccessControlAttribute

Arguments

:: Text
$sel:key:AccessControlAttribute'

-> AccessControlAttributeValue
$sel:value:AccessControlAttribute'

-> AccessControlAttribute

Create a value of AccessControlAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:AccessControlAttribute', accessControlAttribute_key - The name of the attribute associated with your identities in your identity source. This is used to map a specified attribute in your identity source with an attribute in Amazon Web Services SSO.

$sel:value:AccessControlAttribute', accessControlAttribute_value - The value used for mapping a specified attribute to an identity source.

accessControlAttribute_key :: Lens' AccessControlAttribute Text

The name of the attribute associated with your identities in your identity source. This is used to map a specified attribute in your identity source with an attribute in Amazon Web Services SSO.

accessControlAttribute_value :: Lens' AccessControlAttribute AccessControlAttributeValue

The value used for mapping a specified attribute to an identity source.

AccessControlAttributeValue

data AccessControlAttributeValue

The value used for mapping a specified attribute to an identity source.

See: newAccessControlAttributeValue smart constructor.

Constructors

AccessControlAttributeValue'

Fields

source :: NonEmpty Text

The identity source to use when mapping a specified attribute to Amazon Web Services SSO.

newAccessControlAttributeValue

Arguments

:: NonEmpty Text
$sel:source:AccessControlAttributeValue'

-> AccessControlAttributeValue

Create a value of AccessControlAttributeValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:source:AccessControlAttributeValue', accessControlAttributeValue_source - The identity source to use when mapping a specified attribute to Amazon Web Services SSO.

accessControlAttributeValue_source :: Lens' AccessControlAttributeValue (NonEmpty Text)

The identity source to use when mapping a specified attribute to Amazon Web Services SSO.

AccountAssignment

data AccountAssignment

The assignment that indicates a principal's limited access to a specified Amazon Web Services account with a specified permission set.

The term principal here refers to a user or group that is defined in Amazon Web Services SSO.

See: newAccountAssignment smart constructor.

Constructors

AccountAssignment'

Fields

principalId :: Maybe Text

An identifier for an object in Amazon Web Services SSO, such as a user or group. PrincipalIds are GUIDs (For example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about PrincipalIds in Amazon Web Services SSO, see the Amazon Web Services SSO Identity Store API Reference.

principalType :: Maybe PrincipalType

The entity type for which the assignment will be created.

accountId :: Maybe Text

The identifier of the Amazon Web Services account.

permissionSetArn :: Maybe Text

The ARN of the permission set. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

newAccountAssignment :: AccountAssignment

Create a value of AccountAssignment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:principalId:AccountAssignment', accountAssignment_principalId - An identifier for an object in Amazon Web Services SSO, such as a user or group. PrincipalIds are GUIDs (For example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about PrincipalIds in Amazon Web Services SSO, see the Amazon Web Services SSO Identity Store API Reference.

$sel:principalType:AccountAssignment', accountAssignment_principalType - The entity type for which the assignment will be created.

$sel:accountId:AccountAssignment', accountAssignment_accountId - The identifier of the Amazon Web Services account.

$sel:permissionSetArn:AccountAssignment', accountAssignment_permissionSetArn - The ARN of the permission set. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

accountAssignment_principalId :: Lens' AccountAssignment (Maybe Text)

An identifier for an object in Amazon Web Services SSO, such as a user or group. PrincipalIds are GUIDs (For example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about PrincipalIds in Amazon Web Services SSO, see the Amazon Web Services SSO Identity Store API Reference.

accountAssignment_principalType :: Lens' AccountAssignment (Maybe PrincipalType)

The entity type for which the assignment will be created.

accountAssignment_accountId :: Lens' AccountAssignment (Maybe Text)

The identifier of the Amazon Web Services account.

accountAssignment_permissionSetArn :: Lens' AccountAssignment (Maybe Text)

The ARN of the permission set. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

AccountAssignmentOperationStatus

data AccountAssignmentOperationStatus

The status of the creation or deletion operation of an assignment that a principal needs to access an account.

See: newAccountAssignmentOperationStatus smart constructor.

Constructors

AccountAssignmentOperationStatus'

Fields

requestId :: Maybe Text

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

status :: Maybe StatusValues

The status of the permission set provisioning process.

failureReason :: Maybe Text

The message that contains an error or exception in case of an operation failure.

targetId :: Maybe Text

TargetID is an Amazon Web Services account identifier, typically a 10-12 digit string (For example, 123456789012).

principalId :: Maybe Text

An identifier for an object in Amazon Web Services SSO, such as a user or group. PrincipalIds are GUIDs (For example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about PrincipalIds in Amazon Web Services SSO, see the Amazon Web Services SSO Identity Store API Reference.

principalType :: Maybe PrincipalType

The entity type for which the assignment will be created.

targetType :: Maybe TargetType

The entity type for which the assignment will be created.

createdDate :: Maybe POSIX

The date that the permission set was created.

permissionSetArn :: Maybe Text

The ARN of the permission set. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

newAccountAssignmentOperationStatus :: AccountAssignmentOperationStatus

Create a value of AccountAssignmentOperationStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:AccountAssignmentOperationStatus', accountAssignmentOperationStatus_requestId - The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

$sel:status:AccountAssignmentOperationStatus', accountAssignmentOperationStatus_status - The status of the permission set provisioning process.

$sel:failureReason:AccountAssignmentOperationStatus', accountAssignmentOperationStatus_failureReason - The message that contains an error or exception in case of an operation failure.

$sel:targetId:AccountAssignmentOperationStatus', accountAssignmentOperationStatus_targetId - TargetID is an Amazon Web Services account identifier, typically a 10-12 digit string (For example, 123456789012).

$sel:principalId:AccountAssignmentOperationStatus', accountAssignmentOperationStatus_principalId - An identifier for an object in Amazon Web Services SSO, such as a user or group. PrincipalIds are GUIDs (For example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about PrincipalIds in Amazon Web Services SSO, see the Amazon Web Services SSO Identity Store API Reference.

$sel:principalType:AccountAssignmentOperationStatus', accountAssignmentOperationStatus_principalType - The entity type for which the assignment will be created.

$sel:targetType:AccountAssignmentOperationStatus', accountAssignmentOperationStatus_targetType - The entity type for which the assignment will be created.

$sel:createdDate:AccountAssignmentOperationStatus', accountAssignmentOperationStatus_createdDate - The date that the permission set was created.

$sel:permissionSetArn:AccountAssignmentOperationStatus', accountAssignmentOperationStatus_permissionSetArn - The ARN of the permission set. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

accountAssignmentOperationStatus_requestId :: Lens' AccountAssignmentOperationStatus (Maybe Text)

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

accountAssignmentOperationStatus_status :: Lens' AccountAssignmentOperationStatus (Maybe StatusValues)

The status of the permission set provisioning process.

accountAssignmentOperationStatus_failureReason :: Lens' AccountAssignmentOperationStatus (Maybe Text)

The message that contains an error or exception in case of an operation failure.

accountAssignmentOperationStatus_targetId :: Lens' AccountAssignmentOperationStatus (Maybe Text)

TargetID is an Amazon Web Services account identifier, typically a 10-12 digit string (For example, 123456789012).

accountAssignmentOperationStatus_principalId :: Lens' AccountAssignmentOperationStatus (Maybe Text)

An identifier for an object in Amazon Web Services SSO, such as a user or group. PrincipalIds are GUIDs (For example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about PrincipalIds in Amazon Web Services SSO, see the Amazon Web Services SSO Identity Store API Reference.

accountAssignmentOperationStatus_principalType :: Lens' AccountAssignmentOperationStatus (Maybe PrincipalType)

The entity type for which the assignment will be created.

accountAssignmentOperationStatus_targetType :: Lens' AccountAssignmentOperationStatus (Maybe TargetType)

The entity type for which the assignment will be created.

accountAssignmentOperationStatus_createdDate :: Lens' AccountAssignmentOperationStatus (Maybe UTCTime)

The date that the permission set was created.

accountAssignmentOperationStatus_permissionSetArn :: Lens' AccountAssignmentOperationStatus (Maybe Text)

The ARN of the permission set. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

AccountAssignmentOperationStatusMetadata

data AccountAssignmentOperationStatusMetadata

Provides information about the AccountAssignment creation request.

See: newAccountAssignmentOperationStatusMetadata smart constructor.

Constructors

AccountAssignmentOperationStatusMetadata'

Fields

requestId :: Maybe Text

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

status :: Maybe StatusValues

The status of the permission set provisioning process.

createdDate :: Maybe POSIX

The date that the permission set was created.

newAccountAssignmentOperationStatusMetadata :: AccountAssignmentOperationStatusMetadata

Create a value of AccountAssignmentOperationStatusMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:AccountAssignmentOperationStatusMetadata', accountAssignmentOperationStatusMetadata_requestId - The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

$sel:status:AccountAssignmentOperationStatusMetadata', accountAssignmentOperationStatusMetadata_status - The status of the permission set provisioning process.

$sel:createdDate:AccountAssignmentOperationStatusMetadata', accountAssignmentOperationStatusMetadata_createdDate - The date that the permission set was created.

accountAssignmentOperationStatusMetadata_requestId :: Lens' AccountAssignmentOperationStatusMetadata (Maybe Text)

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

accountAssignmentOperationStatusMetadata_status :: Lens' AccountAssignmentOperationStatusMetadata (Maybe StatusValues)

The status of the permission set provisioning process.

accountAssignmentOperationStatusMetadata_createdDate :: Lens' AccountAssignmentOperationStatusMetadata (Maybe UTCTime)

The date that the permission set was created.

AttachedManagedPolicy

data AttachedManagedPolicy

A structure that stores the details of the IAM managed policy.

See: newAttachedManagedPolicy smart constructor.

Constructors

AttachedManagedPolicy'

Fields

arn :: Maybe Text

The ARN of the IAM managed policy. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

name :: Maybe Text

The name of the IAM managed policy.

newAttachedManagedPolicy :: AttachedManagedPolicy

Create a value of AttachedManagedPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:AttachedManagedPolicy', attachedManagedPolicy_arn - The ARN of the IAM managed policy. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

$sel:name:AttachedManagedPolicy', attachedManagedPolicy_name - The name of the IAM managed policy.

attachedManagedPolicy_arn :: Lens' AttachedManagedPolicy (Maybe Text)

The ARN of the IAM managed policy. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

attachedManagedPolicy_name :: Lens' AttachedManagedPolicy (Maybe Text)

The name of the IAM managed policy.

InstanceAccessControlAttributeConfiguration

data InstanceAccessControlAttributeConfiguration

Specifies the attributes to add to your attribute-based access control (ABAC) configuration.

See: newInstanceAccessControlAttributeConfiguration smart constructor.

Constructors

InstanceAccessControlAttributeConfiguration'

Fields

accessControlAttributes :: [AccessControlAttribute]

Lists the attributes that are configured for ABAC in the specified Amazon Web Services SSO instance.

newInstanceAccessControlAttributeConfiguration :: InstanceAccessControlAttributeConfiguration

Create a value of InstanceAccessControlAttributeConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessControlAttributes:InstanceAccessControlAttributeConfiguration', instanceAccessControlAttributeConfiguration_accessControlAttributes - Lists the attributes that are configured for ABAC in the specified Amazon Web Services SSO instance.

instanceAccessControlAttributeConfiguration_accessControlAttributes :: Lens' InstanceAccessControlAttributeConfiguration [AccessControlAttribute]

Lists the attributes that are configured for ABAC in the specified Amazon Web Services SSO instance.

InstanceMetadata

data InstanceMetadata

Provides information about the SSO instance.

See: newInstanceMetadata smart constructor.

Constructors

InstanceMetadata'

Fields

identityStoreId :: Maybe Text

The identifier of the identity store that is connected to the SSO instance.

instanceArn :: Maybe Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

newInstanceMetadata :: InstanceMetadata

Create a value of InstanceMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityStoreId:InstanceMetadata', instanceMetadata_identityStoreId - The identifier of the identity store that is connected to the SSO instance.

$sel:instanceArn:InstanceMetadata', instanceMetadata_instanceArn - The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

instanceMetadata_identityStoreId :: Lens' InstanceMetadata (Maybe Text)

The identifier of the identity store that is connected to the SSO instance.

instanceMetadata_instanceArn :: Lens' InstanceMetadata (Maybe Text)

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

OperationStatusFilter

data OperationStatusFilter

Filters he operation status list based on the passed attribute value.

See: newOperationStatusFilter smart constructor.

Constructors

OperationStatusFilter'

Fields

status :: Maybe StatusValues

Filters the list operations result based on the status attribute.

newOperationStatusFilter :: OperationStatusFilter

Create a value of OperationStatusFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:OperationStatusFilter', operationStatusFilter_status - Filters the list operations result based on the status attribute.

operationStatusFilter_status :: Lens' OperationStatusFilter (Maybe StatusValues)

Filters the list operations result based on the status attribute.

PermissionSet

data PermissionSet

An entity that contains IAM policies.

See: newPermissionSet smart constructor.

Constructors

PermissionSet'

Fields

relayState :: Maybe Text

Used to redirect users within the application during the federation authentication process.

sessionDuration :: Maybe Text

The length of time that the application user sessions are valid for in the ISO-8601 standard.

createdDate :: Maybe POSIX

The date that the permission set was created.

permissionSetArn :: Maybe Text

The ARN of the permission set. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

name :: Maybe Text

The name of the permission set.

description :: Maybe Text

The description of the PermissionSet.

newPermissionSet :: PermissionSet

Create a value of PermissionSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:relayState:PermissionSet', permissionSet_relayState - Used to redirect users within the application during the federation authentication process.

$sel:sessionDuration:PermissionSet', permissionSet_sessionDuration - The length of time that the application user sessions are valid for in the ISO-8601 standard.

$sel:createdDate:PermissionSet', permissionSet_createdDate - The date that the permission set was created.

$sel:permissionSetArn:PermissionSet', permissionSet_permissionSetArn - The ARN of the permission set. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

$sel:name:PermissionSet', permissionSet_name - The name of the permission set.

$sel:description:PermissionSet', permissionSet_description - The description of the PermissionSet.

permissionSet_relayState :: Lens' PermissionSet (Maybe Text)

Used to redirect users within the application during the federation authentication process.

permissionSet_sessionDuration :: Lens' PermissionSet (Maybe Text)

The length of time that the application user sessions are valid for in the ISO-8601 standard.

permissionSet_createdDate :: Lens' PermissionSet (Maybe UTCTime)

The date that the permission set was created.

permissionSet_permissionSetArn :: Lens' PermissionSet (Maybe Text)

The ARN of the permission set. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

permissionSet_name :: Lens' PermissionSet (Maybe Text)

The name of the permission set.

permissionSet_description :: Lens' PermissionSet (Maybe Text)

The description of the PermissionSet.

PermissionSetProvisioningStatus

data PermissionSetProvisioningStatus

A structure that is used to provide the status of the provisioning operation for a specified permission set.

See: newPermissionSetProvisioningStatus smart constructor.

Constructors

PermissionSetProvisioningStatus'

Fields

requestId :: Maybe Text

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

status :: Maybe StatusValues

The status of the permission set provisioning process.

failureReason :: Maybe Text

The message that contains an error or exception in case of an operation failure.

accountId :: Maybe Text

The identifier of the Amazon Web Services account from which to list the assignments.

createdDate :: Maybe POSIX

The date that the permission set was created.

permissionSetArn :: Maybe Text

The ARN of the permission set that is being provisioned. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

newPermissionSetProvisioningStatus :: PermissionSetProvisioningStatus

Create a value of PermissionSetProvisioningStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:PermissionSetProvisioningStatus', permissionSetProvisioningStatus_requestId - The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

$sel:status:PermissionSetProvisioningStatus', permissionSetProvisioningStatus_status - The status of the permission set provisioning process.

$sel:failureReason:PermissionSetProvisioningStatus', permissionSetProvisioningStatus_failureReason - The message that contains an error or exception in case of an operation failure.

$sel:accountId:PermissionSetProvisioningStatus', permissionSetProvisioningStatus_accountId - The identifier of the Amazon Web Services account from which to list the assignments.

$sel:createdDate:PermissionSetProvisioningStatus', permissionSetProvisioningStatus_createdDate - The date that the permission set was created.

$sel:permissionSetArn:PermissionSetProvisioningStatus', permissionSetProvisioningStatus_permissionSetArn - The ARN of the permission set that is being provisioned. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

permissionSetProvisioningStatus_requestId :: Lens' PermissionSetProvisioningStatus (Maybe Text)

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

permissionSetProvisioningStatus_status :: Lens' PermissionSetProvisioningStatus (Maybe StatusValues)

The status of the permission set provisioning process.

permissionSetProvisioningStatus_failureReason :: Lens' PermissionSetProvisioningStatus (Maybe Text)

The message that contains an error or exception in case of an operation failure.

permissionSetProvisioningStatus_accountId :: Lens' PermissionSetProvisioningStatus (Maybe Text)

The identifier of the Amazon Web Services account from which to list the assignments.

permissionSetProvisioningStatus_createdDate :: Lens' PermissionSetProvisioningStatus (Maybe UTCTime)

The date that the permission set was created.

permissionSetProvisioningStatus_permissionSetArn :: Lens' PermissionSetProvisioningStatus (Maybe Text)

The ARN of the permission set that is being provisioned. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

PermissionSetProvisioningStatusMetadata

data PermissionSetProvisioningStatusMetadata

Provides information about the permission set provisioning status.

See: newPermissionSetProvisioningStatusMetadata smart constructor.

Constructors

PermissionSetProvisioningStatusMetadata'

Fields

requestId :: Maybe Text

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

status :: Maybe StatusValues

The status of the permission set provisioning process.

createdDate :: Maybe POSIX

The date that the permission set was created.

newPermissionSetProvisioningStatusMetadata :: PermissionSetProvisioningStatusMetadata

Create a value of PermissionSetProvisioningStatusMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:PermissionSetProvisioningStatusMetadata', permissionSetProvisioningStatusMetadata_requestId - The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

$sel:status:PermissionSetProvisioningStatusMetadata', permissionSetProvisioningStatusMetadata_status - The status of the permission set provisioning process.

$sel:createdDate:PermissionSetProvisioningStatusMetadata', permissionSetProvisioningStatusMetadata_createdDate - The date that the permission set was created.

permissionSetProvisioningStatusMetadata_requestId :: Lens' PermissionSetProvisioningStatusMetadata (Maybe Text)

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

permissionSetProvisioningStatusMetadata_status :: Lens' PermissionSetProvisioningStatusMetadata (Maybe StatusValues)

The status of the permission set provisioning process.

permissionSetProvisioningStatusMetadata_createdDate :: Lens' PermissionSetProvisioningStatusMetadata (Maybe UTCTime)

The date that the permission set was created.

Tag

data Tag

A set of key-value pairs that are used to manage the resource. Tags can only be applied to permission sets and cannot be applied to corresponding roles that Amazon Web Services SSO creates in Amazon Web Services accounts.

See: newTag smart constructor.

Constructors

Tag'

Fields

value :: Maybe Text

The value of the tag.

key :: Maybe Text

The key for the tag.

newTag :: Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:Tag', tag_value - The value of the tag.

$sel:key:Tag', tag_key - The key for the tag.

tag_value :: Lens' Tag (Maybe Text)

The value of the tag.

tag_key :: Lens' Tag (Maybe Text)

The key for the tag.