Network.AWS.SSOAdmin.Types.AccountAssignmentOperationStatus

Description

Documentation

data AccountAssignmentOperationStatus

The status of the creation or deletion operation of an assignment that a principal needs to access an account.

See: newAccountAssignmentOperationStatus smart constructor.

Constructors

AccountAssignmentOperationStatus'

Fields

requestId :: Maybe Text

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

status :: Maybe StatusValues

The status of the permission set provisioning process.

failureReason :: Maybe Text

The message that contains an error or exception in case of an operation failure.

targetId :: Maybe Text

TargetID is an Amazon Web Services account identifier, typically a 10-12 digit string (For example, 123456789012).

principalId :: Maybe Text

An identifier for an object in Amazon Web Services SSO, such as a user or group. PrincipalIds are GUIDs (For example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about PrincipalIds in Amazon Web Services SSO, see the Amazon Web Services SSO Identity Store API Reference.

principalType :: Maybe PrincipalType

The entity type for which the assignment will be created.

targetType :: Maybe TargetType

The entity type for which the assignment will be created.

createdDate :: Maybe POSIX

The date that the permission set was created.

permissionSetArn :: Maybe Text

The ARN of the permission set. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

newAccountAssignmentOperationStatus :: AccountAssignmentOperationStatus

Create a value of AccountAssignmentOperationStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:AccountAssignmentOperationStatus', accountAssignmentOperationStatus_requestId - The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

$sel:status:AccountAssignmentOperationStatus', accountAssignmentOperationStatus_status - The status of the permission set provisioning process.

$sel:failureReason:AccountAssignmentOperationStatus', accountAssignmentOperationStatus_failureReason - The message that contains an error or exception in case of an operation failure.

$sel:targetId:AccountAssignmentOperationStatus', accountAssignmentOperationStatus_targetId - TargetID is an Amazon Web Services account identifier, typically a 10-12 digit string (For example, 123456789012).

$sel:principalId:AccountAssignmentOperationStatus', accountAssignmentOperationStatus_principalId - An identifier for an object in Amazon Web Services SSO, such as a user or group. PrincipalIds are GUIDs (For example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about PrincipalIds in Amazon Web Services SSO, see the Amazon Web Services SSO Identity Store API Reference.

$sel:principalType:AccountAssignmentOperationStatus', accountAssignmentOperationStatus_principalType - The entity type for which the assignment will be created.

$sel:targetType:AccountAssignmentOperationStatus', accountAssignmentOperationStatus_targetType - The entity type for which the assignment will be created.

$sel:createdDate:AccountAssignmentOperationStatus', accountAssignmentOperationStatus_createdDate - The date that the permission set was created.

$sel:permissionSetArn:AccountAssignmentOperationStatus', accountAssignmentOperationStatus_permissionSetArn - The ARN of the permission set. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

accountAssignmentOperationStatus_requestId :: Lens' AccountAssignmentOperationStatus (Maybe Text)

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

accountAssignmentOperationStatus_status :: Lens' AccountAssignmentOperationStatus (Maybe StatusValues)

The status of the permission set provisioning process.

accountAssignmentOperationStatus_failureReason :: Lens' AccountAssignmentOperationStatus (Maybe Text)

The message that contains an error or exception in case of an operation failure.

accountAssignmentOperationStatus_targetId :: Lens' AccountAssignmentOperationStatus (Maybe Text)

TargetID is an Amazon Web Services account identifier, typically a 10-12 digit string (For example, 123456789012).

accountAssignmentOperationStatus_principalId :: Lens' AccountAssignmentOperationStatus (Maybe Text)

An identifier for an object in Amazon Web Services SSO, such as a user or group. PrincipalIds are GUIDs (For example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about PrincipalIds in Amazon Web Services SSO, see the Amazon Web Services SSO Identity Store API Reference.

accountAssignmentOperationStatus_principalType :: Lens' AccountAssignmentOperationStatus (Maybe PrincipalType)

The entity type for which the assignment will be created.

accountAssignmentOperationStatus_targetType :: Lens' AccountAssignmentOperationStatus (Maybe TargetType)

The entity type for which the assignment will be created.

accountAssignmentOperationStatus_createdDate :: Lens' AccountAssignmentOperationStatus (Maybe UTCTime)

The date that the permission set was created.

accountAssignmentOperationStatus_permissionSetArn :: Lens' AccountAssignmentOperationStatus (Maybe Text)

The ARN of the permission set. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.