Network.AWS.SSOAdmin.Types.AccountAssignmentOperationStatusMetadata

Description

Documentation

data AccountAssignmentOperationStatusMetadata

Provides information about the AccountAssignment creation request.

See: newAccountAssignmentOperationStatusMetadata smart constructor.

Constructors

AccountAssignmentOperationStatusMetadata'

Fields

requestId :: Maybe Text

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

status :: Maybe StatusValues

The status of the permission set provisioning process.

createdDate :: Maybe POSIX

The date that the permission set was created.

newAccountAssignmentOperationStatusMetadata :: AccountAssignmentOperationStatusMetadata

Create a value of AccountAssignmentOperationStatusMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:AccountAssignmentOperationStatusMetadata', accountAssignmentOperationStatusMetadata_requestId - The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

$sel:status:AccountAssignmentOperationStatusMetadata', accountAssignmentOperationStatusMetadata_status - The status of the permission set provisioning process.

$sel:createdDate:AccountAssignmentOperationStatusMetadata', accountAssignmentOperationStatusMetadata_createdDate - The date that the permission set was created.

accountAssignmentOperationStatusMetadata_requestId :: Lens' AccountAssignmentOperationStatusMetadata (Maybe Text)

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

accountAssignmentOperationStatusMetadata_status :: Lens' AccountAssignmentOperationStatusMetadata (Maybe StatusValues)

The status of the permission set provisioning process.

accountAssignmentOperationStatusMetadata_createdDate :: Lens' AccountAssignmentOperationStatusMetadata (Maybe UTCTime)

The date that the permission set was created.