Network.AWS.SSOOIDC.RegisterClient

Description

Registers a client with AWS SSO. This allows clients to initiate device authorization. The output should be persisted for reuse through many authentication requests.

Creating a Request

data RegisterClient

See: newRegisterClient smart constructor.

Constructors

RegisterClient'

Fields

scopes :: Maybe [Text]

The list of scopes that are defined by the client. Upon authorization, this list is used to restrict permissions when granting an access token.

clientName :: Text

The friendly name of the client.

clientType :: Text

The type of client. The service supports only public as a client type. Anything other than public will be rejected by the service.

newRegisterClient

Arguments

:: Text
$sel:clientName:RegisterClient'

-> Text
$sel:clientType:RegisterClient'

-> RegisterClient

Create a value of RegisterClient with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scopes:RegisterClient', registerClient_scopes - The list of scopes that are defined by the client. Upon authorization, this list is used to restrict permissions when granting an access token.

$sel:clientName:RegisterClient', registerClient_clientName - The friendly name of the client.

$sel:clientType:RegisterClient', registerClient_clientType - The type of client. The service supports only public as a client type. Anything other than public will be rejected by the service.

Request Lenses

registerClient_scopes :: Lens' RegisterClient (Maybe [Text])

The list of scopes that are defined by the client. Upon authorization, this list is used to restrict permissions when granting an access token.

registerClient_clientName :: Lens' RegisterClient Text

The friendly name of the client.

registerClient_clientType :: Lens' RegisterClient Text

The type of client. The service supports only public as a client type. Anything other than public will be rejected by the service.

Destructuring the Response

data RegisterClientResponse

See: newRegisterClientResponse smart constructor.

Constructors

RegisterClientResponse'

Fields

clientId :: Maybe Text

The unique identifier string for each client. This client uses this identifier to get authenticated by the service in subsequent calls.

clientSecret :: Maybe Text

A secret string generated for the client. The client will use this string to get authenticated by the service in subsequent calls.

clientIdIssuedAt :: Maybe Integer

Indicates the time at which the clientId and clientSecret were issued.

clientSecretExpiresAt :: Maybe Integer

Indicates the time at which the clientId and clientSecret will become invalid.

tokenEndpoint :: Maybe Text

The endpoint where the client can get an access token.

authorizationEndpoint :: Maybe Text

The endpoint where the client can request authorization.

httpStatus :: Int

The response's http status code.

newRegisterClientResponse

Arguments

:: Int
$sel:httpStatus:RegisterClientResponse'

-> RegisterClientResponse

Create a value of RegisterClientResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientId:RegisterClientResponse', registerClientResponse_clientId - The unique identifier string for each client. This client uses this identifier to get authenticated by the service in subsequent calls.

$sel:clientSecret:RegisterClientResponse', registerClientResponse_clientSecret - A secret string generated for the client. The client will use this string to get authenticated by the service in subsequent calls.

$sel:clientIdIssuedAt:RegisterClientResponse', registerClientResponse_clientIdIssuedAt - Indicates the time at which the clientId and clientSecret were issued.

$sel:clientSecretExpiresAt:RegisterClientResponse', registerClientResponse_clientSecretExpiresAt - Indicates the time at which the clientId and clientSecret will become invalid.

$sel:tokenEndpoint:RegisterClientResponse', registerClientResponse_tokenEndpoint - The endpoint where the client can get an access token.

$sel:authorizationEndpoint:RegisterClientResponse', registerClientResponse_authorizationEndpoint - The endpoint where the client can request authorization.

$sel:httpStatus:RegisterClientResponse', registerClientResponse_httpStatus - The response's http status code.

Response Lenses

registerClientResponse_clientId :: Lens' RegisterClientResponse (Maybe Text)

The unique identifier string for each client. This client uses this identifier to get authenticated by the service in subsequent calls.

registerClientResponse_clientSecret :: Lens' RegisterClientResponse (Maybe Text)

A secret string generated for the client. The client will use this string to get authenticated by the service in subsequent calls.

registerClientResponse_clientIdIssuedAt :: Lens' RegisterClientResponse (Maybe Integer)

Indicates the time at which the clientId and clientSecret were issued.

registerClientResponse_clientSecretExpiresAt :: Lens' RegisterClientResponse (Maybe Integer)

Indicates the time at which the clientId and clientSecret will become invalid.

registerClientResponse_tokenEndpoint :: Lens' RegisterClientResponse (Maybe Text)

The endpoint where the client can get an access token.

registerClientResponse_authorizationEndpoint :: Lens' RegisterClientResponse (Maybe Text)

The endpoint where the client can request authorization.

registerClientResponse_httpStatus :: Lens' RegisterClientResponse Int

The response's http status code.