Network.AWS.SSOOIDC.StartDeviceAuthorization

Description

Initiates device authorization by requesting a pair of verification codes from the authorization service.

Creating a Request

data StartDeviceAuthorization

See: newStartDeviceAuthorization smart constructor.

Constructors

StartDeviceAuthorization'

Fields

clientId :: Text

The unique identifier string for the client that is registered with AWS SSO. This value should come from the persisted result of the RegisterClient API operation.

clientSecret :: Text

A secret string that is generated for the client. This value should come from the persisted result of the RegisterClient API operation.

startUrl :: Text

The URL for the AWS SSO user portal. For more information, see Using the User Portal in the AWS Single Sign-On User Guide.

newStartDeviceAuthorization

Arguments

:: Text
$sel:clientId:StartDeviceAuthorization'

-> Text
$sel:clientSecret:StartDeviceAuthorization'

-> Text
$sel:startUrl:StartDeviceAuthorization'

-> StartDeviceAuthorization

Create a value of StartDeviceAuthorization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientId:StartDeviceAuthorization', startDeviceAuthorization_clientId - The unique identifier string for the client that is registered with AWS SSO. This value should come from the persisted result of the RegisterClient API operation.

$sel:clientSecret:StartDeviceAuthorization', startDeviceAuthorization_clientSecret - A secret string that is generated for the client. This value should come from the persisted result of the RegisterClient API operation.

$sel:startUrl:StartDeviceAuthorization', startDeviceAuthorization_startUrl - The URL for the AWS SSO user portal. For more information, see Using the User Portal in the AWS Single Sign-On User Guide.

Request Lenses

startDeviceAuthorization_clientId :: Lens' StartDeviceAuthorization Text

The unique identifier string for the client that is registered with AWS SSO. This value should come from the persisted result of the RegisterClient API operation.

startDeviceAuthorization_clientSecret :: Lens' StartDeviceAuthorization Text

A secret string that is generated for the client. This value should come from the persisted result of the RegisterClient API operation.

startDeviceAuthorization_startUrl :: Lens' StartDeviceAuthorization Text

The URL for the AWS SSO user portal. For more information, see Using the User Portal in the AWS Single Sign-On User Guide.

Destructuring the Response

data StartDeviceAuthorizationResponse

See: newStartDeviceAuthorizationResponse smart constructor.

Constructors

StartDeviceAuthorizationResponse'

Fields

userCode :: Maybe Text

A one-time user verification code. This is needed to authorize an in-use device.

interval :: Maybe Int

Indicates the number of seconds the client must wait between attempts when polling for a session.

expiresIn :: Maybe Int

Indicates the number of seconds in which the verification code will become invalid.

verificationUri :: Maybe Text

The URI of the verification page that takes the userCode to authorize the device.

deviceCode :: Maybe Text

The short-lived code that is used by the device when polling for a session token.

verificationUriComplete :: Maybe Text

An alternate URL that the client can use to automatically launch a browser. This process skips the manual step in which the user visits the verification page and enters their code.

httpStatus :: Int

The response's http status code.

newStartDeviceAuthorizationResponse

Arguments

:: Int
$sel:httpStatus:StartDeviceAuthorizationResponse'

-> StartDeviceAuthorizationResponse

Create a value of StartDeviceAuthorizationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userCode:StartDeviceAuthorizationResponse', startDeviceAuthorizationResponse_userCode - A one-time user verification code. This is needed to authorize an in-use device.

$sel:interval:StartDeviceAuthorizationResponse', startDeviceAuthorizationResponse_interval - Indicates the number of seconds the client must wait between attempts when polling for a session.

$sel:expiresIn:StartDeviceAuthorizationResponse', startDeviceAuthorizationResponse_expiresIn - Indicates the number of seconds in which the verification code will become invalid.

$sel:verificationUri:StartDeviceAuthorizationResponse', startDeviceAuthorizationResponse_verificationUri - The URI of the verification page that takes the userCode to authorize the device.

$sel:deviceCode:StartDeviceAuthorizationResponse', startDeviceAuthorizationResponse_deviceCode - The short-lived code that is used by the device when polling for a session token.

$sel:verificationUriComplete:StartDeviceAuthorizationResponse', startDeviceAuthorizationResponse_verificationUriComplete - An alternate URL that the client can use to automatically launch a browser. This process skips the manual step in which the user visits the verification page and enters their code.

$sel:httpStatus:StartDeviceAuthorizationResponse', startDeviceAuthorizationResponse_httpStatus - The response's http status code.

Response Lenses

startDeviceAuthorizationResponse_userCode :: Lens' StartDeviceAuthorizationResponse (Maybe Text)

A one-time user verification code. This is needed to authorize an in-use device.

startDeviceAuthorizationResponse_interval :: Lens' StartDeviceAuthorizationResponse (Maybe Int)

Indicates the number of seconds the client must wait between attempts when polling for a session.

startDeviceAuthorizationResponse_expiresIn :: Lens' StartDeviceAuthorizationResponse (Maybe Int)

Indicates the number of seconds in which the verification code will become invalid.

startDeviceAuthorizationResponse_verificationUri :: Lens' StartDeviceAuthorizationResponse (Maybe Text)

The URI of the verification page that takes the userCode to authorize the device.

startDeviceAuthorizationResponse_deviceCode :: Lens' StartDeviceAuthorizationResponse (Maybe Text)

The short-lived code that is used by the device when polling for a session token.

startDeviceAuthorizationResponse_verificationUriComplete :: Lens' StartDeviceAuthorizationResponse (Maybe Text)

An alternate URL that the client can use to automatically launch a browser. This process skips the manual step in which the user visits the verification page and enters their code.

startDeviceAuthorizationResponse_httpStatus :: Lens' StartDeviceAuthorizationResponse Int

The response's http status code.