Network.AWS.STS.DecodeAuthorizationMessage

Description

Decodes additional information about the authorization status of a request from an encoded message returned in response to an Amazon Web Services request.

For example, if a user is not authorized to perform an operation that he or she has requested, the request returns a Client.UnauthorizedOperation response (an HTTP 403 response). Some Amazon Web Services operations additionally return an encoded message that can provide details about this authorization failure.

Only certain Amazon Web Services operations return an encoded authorization message. The documentation for an individual operation indicates whether that operation returns an encoded message in addition to returning an HTTP code.

The message is encoded because the details of the authorization status can constitute privileged information that the user who requested the operation should not see. To decode an authorization status message, a user must be granted permissions via an IAM policy to request the DecodeAuthorizationMessage (sts:DecodeAuthorizationMessage) action.

The decoded message includes the following type of information:

Whether the request was denied due to an explicit deny or due to the absence of an explicit allow. For more information, see Determining Whether a Request is Allowed or Denied in the IAM User Guide.

The principal who made the request.

The requested action.

The requested resource.

The values of condition keys in the context of the user's request.

Creating a Request

data DecodeAuthorizationMessage

See: newDecodeAuthorizationMessage smart constructor.

Constructors

DecodeAuthorizationMessage'

Fields

encodedMessage :: Text

The encoded message that was returned with the response.

newDecodeAuthorizationMessage

Arguments

:: Text
$sel:encodedMessage:DecodeAuthorizationMessage'

-> DecodeAuthorizationMessage

Create a value of DecodeAuthorizationMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encodedMessage:DecodeAuthorizationMessage', decodeAuthorizationMessage_encodedMessage - The encoded message that was returned with the response.

Request Lenses

decodeAuthorizationMessage_encodedMessage :: Lens' DecodeAuthorizationMessage Text

The encoded message that was returned with the response.

Destructuring the Response

data DecodeAuthorizationMessageResponse

A document that contains additional information about the authorization status of a request from an encoded message that is returned in response to an Amazon Web Services request.

See: newDecodeAuthorizationMessageResponse smart constructor.

Constructors

DecodeAuthorizationMessageResponse'

Fields

decodedMessage :: Maybe Text

An XML document that contains the decoded message.

httpStatus :: Int

The response's http status code.

newDecodeAuthorizationMessageResponse

Arguments

:: Int
$sel:httpStatus:DecodeAuthorizationMessageResponse'

-> DecodeAuthorizationMessageResponse

Create a value of DecodeAuthorizationMessageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:decodedMessage:DecodeAuthorizationMessageResponse', decodeAuthorizationMessageResponse_decodedMessage - An XML document that contains the decoded message.

$sel:httpStatus:DecodeAuthorizationMessageResponse', decodeAuthorizationMessageResponse_httpStatus - The response's http status code.

Response Lenses

decodeAuthorizationMessageResponse_decodedMessage :: Lens' DecodeAuthorizationMessageResponse (Maybe Text)

An XML document that contains the decoded message.

decodeAuthorizationMessageResponse_httpStatus :: Lens' DecodeAuthorizationMessageResponse Int

The response's http status code.