Network.AWS.STS.GetCallerIdentity

Description

Returns details about the IAM user or role whose credentials are used to call the operation.

No permissions are required to perform this operation. If an administrator adds a policy to your IAM user or role that explicitly denies access to the sts:GetCallerIdentity action, you can still perform this operation. Permissions are not required because the same information is returned when an IAM user or role is denied access. To view an example response, see I Am Not Authorized to Perform: iam:DeleteVirtualMFADevice in the IAM User Guide.

Creating a Request

data GetCallerIdentity

See: newGetCallerIdentity smart constructor.

Constructors

GetCallerIdentity'

newGetCallerIdentity :: GetCallerIdentity

Create a value of GetCallerIdentity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetCallerIdentityResponse

Contains the response to a successful GetCallerIdentity request, including information about the entity making the request.

See: newGetCallerIdentityResponse smart constructor.

Constructors

GetCallerIdentityResponse'

Fields

arn :: Maybe Text

The Amazon Web Services ARN associated with the calling entity.

userId :: Maybe Text

The unique identifier of the calling entity. The exact value depends on the type of entity that is making the call. The values returned are those listed in the aws:userid column in the Principal table found on the Policy Variables reference page in the /IAM User Guide/.

account :: Maybe Text

The Amazon Web Services account ID number of the account that owns or contains the calling entity.

httpStatus :: Int

The response's http status code.

newGetCallerIdentityResponse

Arguments

:: Int
$sel:httpStatus:GetCallerIdentityResponse'

-> GetCallerIdentityResponse

Create a value of GetCallerIdentityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:GetCallerIdentityResponse', getCallerIdentityResponse_arn - The Amazon Web Services ARN associated with the calling entity.

$sel:userId:GetCallerIdentityResponse', getCallerIdentityResponse_userId - The unique identifier of the calling entity. The exact value depends on the type of entity that is making the call. The values returned are those listed in the aws:userid column in the Principal table found on the Policy Variables reference page in the /IAM User Guide/.

$sel:account:GetCallerIdentityResponse', getCallerIdentityResponse_account - The Amazon Web Services account ID number of the account that owns or contains the calling entity.

$sel:httpStatus:GetCallerIdentityResponse', getCallerIdentityResponse_httpStatus - The response's http status code.

Response Lenses

getCallerIdentityResponse_arn :: Lens' GetCallerIdentityResponse (Maybe Text)

The Amazon Web Services ARN associated with the calling entity.

getCallerIdentityResponse_userId :: Lens' GetCallerIdentityResponse (Maybe Text)

The unique identifier of the calling entity. The exact value depends on the type of entity that is making the call. The values returned are those listed in the aws:userid column in the Principal table found on the Policy Variables reference page in the /IAM User Guide/.

getCallerIdentityResponse_account :: Lens' GetCallerIdentityResponse (Maybe Text)

The Amazon Web Services account ID number of the account that owns or contains the calling entity.

getCallerIdentityResponse_httpStatus :: Lens' GetCallerIdentityResponse Int

The response's http status code.