Network.AWS.STS.V2011_06_15

Description

The AWS Security Token Service (STS) is a web service that enables you to request temporary, limited-privilege credentials for AWS Identity and Access Management (IAM) users or for users that you authenticate (federated users).

Documentation

assumeRole

Arguments

:: Text
_arrRoleArn

-> Text
_arrRoleSessionName

-> AssumeRole

Minimum specification for a AssumeRole request.

data AssumeRole

Constructors

AssumeRole

Fields

_arrRoleArn :: Text

The Amazon Resource Name (ARN) of the role that the caller is assuming.

_arrRoleSessionName :: Text

An identifier for the assumed role session. The session name is included as part of the AssumedRoleUser.

_arrExternalId :: Maybe Text

A unique identifier that is used by third parties to assume a role in their customers' accounts. For each role that the third party can assume, they should instruct their customers to create a role with the external ID that the third party generated. Each time the third party assumes the role, they must pass the customer's external ID. The external ID is useful in order to help third parties bind a role to the customer who created it. For more information about the external ID, see About the External ID in Using Temporary Security Credentials.

_arrDurationSeconds :: Maybe Integer

The duration, in seconds, of the role session. The value can range from 900 seconds (15 minutes) to 3600 seconds (1 hour). By default, the value is set to 3600 seconds.

_arrSerialNumber :: Maybe Text

The identification number of the MFA device that is associated with the user who is making the AssumeRole call. Specify this value if the trust policy of the role being assumed includes a condition that requires MFA authentication. The value is either the serial number for a hardware device (such as GAHT12345678) or an Amazon Resource Name (ARN) for a virtual device (such as arn:aws:iam::123456789012:mfa/user).

_arrPolicy :: Maybe Text

An IAM policy in JSON format. The policy parameter is optional. If you pass a policy, the temporary security credentials that are returned by the operation have the permissions that are allowed by both the access policy of the role that is being assumed, and the policy that you pass. This gives you a way to further restrict the permissions for the resulting temporary security credentials. You cannot use the passed policy to grant permissions that are in excess of those allowed by the access policy of the role that is being assumed. For more information, see Permissions for AssumeRole in Using Temporary Security Credentials.

_arrTokenCode :: Maybe Text

The value provided by the MFA device, if the trust policy of the role being assumed requires MFA (that is, if the policy includes a condition that tests for MFA). If the role being assumed requires MFA and if the TokenCode value is missing or expired, the AssumeRole call returns an "access denied" error.

arrTokenCode :: Lens' AssumeRole (Maybe Text)

arrSerialNumber :: Lens' AssumeRole (Maybe Text)

arrRoleSessionName :: Lens' AssumeRole Text

arrRoleArn :: Lens' AssumeRole Text

arrPolicy :: Lens' AssumeRole (Maybe Text)

arrExternalId :: Lens' AssumeRole (Maybe Text)

arrDurationSeconds :: Lens' AssumeRole (Maybe Integer)

data AssumeRoleResponse

Constructors

AssumeRoleResponse

Fields

_arsAssumedRoleUser :: Maybe AssumedRoleUser

The Amazon Resource Name (ARN) and the assumed role ID, which are identifiers that you can use to refer to the resulting temporary security credentials. For example, you can reference these credentials as a principal in a resource-based policy by using the ARN or assumed role ID. The ARN and ID include the RoleSessionName that you specified when you called AssumeRole.

_arsCredentials :: Maybe Credentials

The temporary security credentials, which include an access key ID, a secret access key, and a security (or session) token.

_arsPackedPolicySize :: Maybe Integer

A percentage value that indicates the size of the policy in packed form. The service rejects any policy with a packed size greater than 100 percent, which means the policy exceeded the allowed space.

arsPackedPolicySize :: Lens' AssumeRoleResponse (Maybe Integer)

arsCredentials :: Lens' AssumeRoleResponse (Maybe Credentials)

arsAssumedRoleUser :: Lens' AssumeRoleResponse (Maybe AssumedRoleUser)

assumeRoleWithSAML

Arguments

:: Text
_arwsamlrPrincipalArn

-> Text
_arwsamlrRoleArn

-> Text
_arwsamlrSAMLAssertion

-> AssumeRoleWithSAML

Minimum specification for a AssumeRoleWithSAML request.

data AssumeRoleWithSAML

Constructors

AssumeRoleWithSAML

Fields

_arwsamlrPrincipalArn :: Text

The Amazon Resource Name (ARN) of the SAML provider in IAM that describes the IdP.

_arwsamlrRoleArn :: Text

The Amazon Resource Name (ARN) of the role that the caller is assuming.

_arwsamlrSAMLAssertion :: Text

The base-64 encoded SAML authentication response provided by the IdP. For more information, see Configuring a Relying Party and Adding Claims in the Using IAM guide.

_arwsamlrDurationSeconds :: Maybe Integer

The duration, in seconds, of the role session. The value can range from 900 seconds (15 minutes) to 3600 seconds (1 hour). By default, the value is set to 3600 seconds. An expiration can also be specified in the SAML authentication response's NotOnOrAfter value. The actual expiration time is whichever value is shorter. The maximum duration for a session is 1 hour, and the minimum duration is 15 minutes, even if values outside this range are specified.

_arwsamlrPolicy :: Maybe Text

An IAM policy in JSON format. The policy parameter is optional. If you pass a policy, the temporary security credentials that are returned by the operation have the permissions that are allowed by both the access policy of the role that is being assumed, and the policy that you pass. This gives you a way to further restrict the permissions for the resulting temporary security credentials. You cannot use the passed policy to grant permissions that are in excess of those allowed by the access policy of the role that is being assumed. For more information, see Permissions for AssumeRoleWithSAML in Using Temporary Security Credentials. The policy must be 2048 bytes or shorter, and its packed size must be less than 450 bytes.

arwsamlrSAMLAssertion :: Lens' AssumeRoleWithSAML Text

arwsamlrRoleArn :: Lens' AssumeRoleWithSAML Text

arwsamlrPrincipalArn :: Lens' AssumeRoleWithSAML Text

arwsamlrPolicy :: Lens' AssumeRoleWithSAML (Maybe Text)

arwsamlrDurationSeconds :: Lens' AssumeRoleWithSAML (Maybe Integer)

data AssumeRoleWithSAMLResponse

Constructors

AssumeRoleWithSAMLResponse

Fields

_arwsamlsAssumedRoleUser :: Maybe AssumedRoleUser

The identifiers for the temporary security credentials that the operation returns.

_arwsamlsAudience :: Maybe Text

_arwsamlsCredentials :: Maybe Credentials

AWS credentials for API authentication.

_arwsamlsIssuer :: Maybe Text

_arwsamlsNameQualifier :: Maybe Text

_arwsamlsPackedPolicySize :: Maybe Integer

A percentage value that indicates the size of the policy in packed form. The service rejects any policy with a packed size greater than 100 percent, which means the policy exceeded the allowed space.

_arwsamlsSubject :: Maybe Text

_arwsamlsSubjectType :: Maybe Text

arwsamlsSubjectType :: Lens' AssumeRoleWithSAMLResponse (Maybe Text)

arwsamlsSubject :: Lens' AssumeRoleWithSAMLResponse (Maybe Text)

arwsamlsPackedPolicySize :: Lens' AssumeRoleWithSAMLResponse (Maybe Integer)

arwsamlsNameQualifier :: Lens' AssumeRoleWithSAMLResponse (Maybe Text)

arwsamlsIssuer :: Lens' AssumeRoleWithSAMLResponse (Maybe Text)

arwsamlsCredentials :: Lens' AssumeRoleWithSAMLResponse (Maybe Credentials)

arwsamlsAudience :: Lens' AssumeRoleWithSAMLResponse (Maybe Text)

arwsamlsAssumedRoleUser :: Lens' AssumeRoleWithSAMLResponse (Maybe AssumedRoleUser)

assumeRoleWithWebIdentity

Arguments

:: Text
_arwwirRoleArn

-> Text
_arwwirWebIdentityToken

-> Text
_arwwirRoleSessionName

-> AssumeRoleWithWebIdentity

Minimum specification for a AssumeRoleWithWebIdentity request.

data AssumeRoleWithWebIdentity

Constructors

AssumeRoleWithWebIdentity

Fields

_arwwirRoleArn :: Text

The Amazon Resource Name (ARN) of the role that the caller is assuming.

_arwwirWebIdentityToken :: Text

The OAuth 2.0 access token or OpenID Connect ID token that is provided by the identity provider. Your application must get this token by authenticating the user who is using your application with a web identity provider before the application makes an AssumeRoleWithWebIdentity call.

_arwwirRoleSessionName :: Text

An identifier for the assumed role session. Typically, you pass the name or identifier that is associated with the user who is using your application. That way, the temporary security credentials that your application will use are associated with that user. This session name is included as part of the ARN and assumed role ID in the AssumedRoleUser response element.

_arwwirDurationSeconds :: Maybe Integer

The duration, in seconds, of the role session. The value can range from 900 seconds (15 minutes) to 3600 seconds (1 hour). By default, the value is set to 3600 seconds.

_arwwirPolicy :: Maybe Text

An IAM policy in JSON format. The policy parameter is optional. If you pass a policy, the temporary security credentials that are returned by the operation have the permissions that are allowed by both the access policy of the role that is being assumed, and the policy that you pass. This gives you a way to further restrict the permissions for the resulting temporary security credentials. You cannot use the passed policy to grant permissions that are in excess of those allowed by the access policy of the role that is being assumed. For more information, see Permissions for AssumeRoleWithWebIdentity in Using Temporary Security Credentials.

_arwwirProviderId :: Maybe Text

The fully-qualified host component of the domain name of the identity provider. Specify this value only for OAuth access tokens. Do not specify this value for OpenID Connect ID tokens, such as accounts.google.com. Do not include URL schemes and port numbers. Currently, www.amazon.com and graph.facebook.com are supported.

arwwirWebIdentityToken :: Lens' AssumeRoleWithWebIdentity Text

arwwirRoleSessionName :: Lens' AssumeRoleWithWebIdentity Text

arwwirRoleArn :: Lens' AssumeRoleWithWebIdentity Text

arwwirProviderId :: Lens' AssumeRoleWithWebIdentity (Maybe Text)

arwwirPolicy :: Lens' AssumeRoleWithWebIdentity (Maybe Text)

arwwirDurationSeconds :: Lens' AssumeRoleWithWebIdentity (Maybe Integer)

data AssumeRoleWithWebIdentityResponse

Constructors

AssumeRoleWithWebIdentityResponse

Fields

_arwwisAssumedRoleUser :: Maybe AssumedRoleUser

The Amazon Resource Name (ARN) and the assumed role ID, which are identifiers that you can use to refer to the resulting temporary security credentials. For example, you can reference these credentials as a principal in a resource-based policy by using the ARN or assumed role ID. The ARN and ID include the RoleSessionName that you specified when you called AssumeRole.

_arwwisAudience :: Maybe Text

The intended audience of the web identity token. This is traditionally the client identifier issued to the application that requested the web identity token.

_arwwisCredentials :: Maybe Credentials

The temporary security credentials, which include an access key ID, a secret access key, and a security token.

_arwwisProvider :: Maybe Text

The issuing authority of the web identity token presented. For OpenID Connect ID Tokens this contains the value of the iss field. For OAuth 2.0 Access Tokens, this contains the value of the ProviderId parameter that was passed in the AssumeRoleWithWebIdentity request.

_arwwisPackedPolicySize :: Maybe Integer

A percentage value that indicates the size of the policy in packed form. The service rejects any policy with a packed size greater than 100 percent, which means the policy exceeded the allowed space.

_arwwisSubjectFromWebIdentityToken :: Maybe Text

The unique user identifier that is returned by the identity provider. This identifier is associated with the WebIdentityToken that was submitted with the AssumeRoleWithWebIdentity call. The identifier is typically unique to the user and the application that acquired the WebIdentityToken (pairwise identifier). If an OpenID Connect ID token was submitted in the WebIdentityToken, this value is returned by the identity provider as the token's sub (Subject) claim.

arwwisSubjectFromWebIdentityToken :: Lens' AssumeRoleWithWebIdentityResponse (Maybe Text)

arwwisProvider :: Lens' AssumeRoleWithWebIdentityResponse (Maybe Text)

arwwisPackedPolicySize :: Lens' AssumeRoleWithWebIdentityResponse (Maybe Integer)

arwwisCredentials :: Lens' AssumeRoleWithWebIdentityResponse (Maybe Credentials)

arwwisAudience :: Lens' AssumeRoleWithWebIdentityResponse (Maybe Text)

arwwisAssumedRoleUser :: Lens' AssumeRoleWithWebIdentityResponse (Maybe AssumedRoleUser)

data DecodeAuthorizationMessage

Constructors

DecodeAuthorizationMessage

Fields

_damrEncodedMessage :: Text

The encoded message that was returned with the response.

damrEncodedMessage :: Iso' DecodeAuthorizationMessage Text

data DecodeAuthorizationMessageResponse

Constructors

DecodeAuthorizationMessageResponse

Fields

_damsDecodedMessage :: Maybe Text

An XML document that contains the decoded message. For more information, see DecodeAuthorizationMessage.

damsDecodedMessage :: Iso' DecodeAuthorizationMessageResponse (Maybe Text)

getFederationToken

Arguments

:: Text
_gftrName

-> GetFederationToken

Minimum specification for a GetFederationToken request.

data GetFederationToken

Constructors

GetFederationToken

Fields

_gftrName :: Text

The name of the federated user. The name is used as an identifier for the temporary security credentials (such as Bob). For example, you can reference the federated user name in a resource-based policy, such as in an Amazon S3 bucket policy.

_gftrDurationSeconds :: Maybe Integer

The duration, in seconds, that the session should last. Acceptable durations for federation sessions range from 900 seconds (15 minutes) to 129600 seconds (36 hours), with 43200 seconds (12 hours) as the default. Sessions obtained using AWS account (root) credentials are restricted to a maximum of 3600 seconds (one hour). If the specified duration is longer than one hour, the session obtained by using AWS account (root) credentials defaults to one hour.

_gftrPolicy :: Maybe Text

An IAM policy in JSON format that is passed with the GetFederationToken call and evaluated along with the policy or policies that are attached to the IAM user whose credentials are used to call GetFederationToken. The passed policy is used to scope down the permissions that are available to the IAM user, by allowing only a subset of the permissions that are granted to the IAM user. The passed policy cannot grant more permissions than those granted to the IAM user. The final permissions for the federated user are the most restrictive set based on the intersection of the passed policy and the IAM user policy. If you do not pass a policy, the resulting temporary security credentials have no effective permissions. The only exception is when the temporary security credentials are used to access a resource that has a resource-based policy that specifically allows the federated user to access the resource. For more information about how permissions work, see Permissions for GetFederationToken in Using Temporary Security Credentials.

gftrPolicy :: Lens' GetFederationToken (Maybe Text)

gftrName :: Lens' GetFederationToken Text

gftrDurationSeconds :: Lens' GetFederationToken (Maybe Integer)

data GetFederationTokenResponse

Constructors

GetFederationTokenResponse

Fields

_gftsCredentials :: Maybe Credentials

Credentials for the service API authentication.

_gftsFederatedUser :: Maybe FederatedUser

Identifiers for the federated user associated with the credentials (such as arn:aws:sts::123456789012:federated-user/Bob or 123456789012:Bob). You can use the federated user's ARN in your resource-based policies, such as an Amazon S3 bucket policy.

_gftsPackedPolicySize :: Maybe Integer

A percentage value indicating the size of the policy in packed form. The service rejects policies for which the packed size is greater than 100 percent of the allowed value.

gftsPackedPolicySize :: Lens' GetFederationTokenResponse (Maybe Integer)

gftsFederatedUser :: Lens' GetFederationTokenResponse (Maybe FederatedUser)

gftsCredentials :: Lens' GetFederationTokenResponse (Maybe Credentials)

getSessionToken :: GetSessionToken

Minimum specification for a GetSessionToken request.

data GetSessionToken

Constructors

GetSessionToken

Fields

_gstrDurationSeconds :: Maybe Integer

The duration, in seconds, that the credentials should remain valid. Acceptable durations for IAM user sessions range from 900 seconds (15 minutes) to 129600 seconds (36 hours), with 43200 seconds (12 hours) as the default. Sessions for AWS account owners are restricted to a maximum of 3600 seconds (one hour). If the duration is longer than one hour, the session for AWS account owners defaults to one hour.

_gstrSerialNumber :: Maybe Text

The identification number of the MFA device that is associated with the IAM user who is making the GetSessionToken call. Specify this value if the IAM user has a policy that requires MFA authentication. The value is either the serial number for a hardware device (such as GAHT12345678) or an Amazon Resource Name (ARN) for a virtual device (such as arn:aws:iam::123456789012:mfa/user). You can find the device for an IAM user by going to the AWS Management Console and viewing the user's security credentials.

_gstrTokenCode :: Maybe Text

The value provided by the MFA device, if MFA is required. If any policy requires the IAM user to submit an MFA code, specify this value. If MFA authentication is required, and the user does not provide a code when requesting a set of temporary security credentials, the user will receive an "access denied" response when requesting resources that require MFA authentication.

gstrTokenCode :: Lens' GetSessionToken (Maybe Text)

gstrSerialNumber :: Lens' GetSessionToken (Maybe Text)

gstrDurationSeconds :: Lens' GetSessionToken (Maybe Integer)

data GetSessionTokenResponse

Constructors

GetSessionTokenResponse

Fields

_gstsCredentials :: Maybe Credentials

The session credentials for API authentication.

gstsCredentials :: Iso' GetSessionTokenResponse (Maybe Credentials)

data STS

Supported version (2011-06-15) of the AWS Security Token Service service.

xmlOptions :: Tagged a XMLOptions

data AssumedRoleUser

The Amazon Resource Name (ARN) and the assumed role ID, which are identifiers that you can use to refer to the resulting temporary security credentials. For example, you can reference these credentials as a principal in a resource-based policy by using the ARN or assumed role ID. The ARN and ID include the RoleSessionName that you specified when you called AssumeRole.

Constructors

AssumedRoleUser

Fields

_aruArn :: Text

The ARN of the temporary security credentials that are returned from the AssumeRole action. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in Using IAM.

_aruAssumedRoleId :: Text

A unique identifier that contains the role ID and the role session name of the role that is being assumed. The role ID is generated by AWS when the role is created.

data Credentials

The temporary security credentials, which include an access key ID, a secret access key, and a security (or session) token.

Constructors

Credentials

Fields

_cSecretAccessKey :: Text

The secret access key that can be used to sign requests.

_cSessionToken :: Text

The token that users must pass to the service API to use the temporary credentials.

_cExpiration :: ISO8601

The date on which the current credentials expire.

_cAccessKeyId :: Text

The access key ID that identifies the temporary security credentials.

data FederatedUser

Identifiers for the federated user associated with the credentials (such as arn:aws:sts::123456789012:federated-user/Bob or 123456789012:Bob). You can use the federated user's ARN in your resource-based policies, such as an Amazon S3 bucket policy.

Constructors

FederatedUser

Fields

_fuArn :: Text

The ARN that specifies the federated user that is associated with the credentials. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in Using IAM.

_fuFederatedUserId :: Text

The string that identifies the federated user associated with the credentials, similar to the unique ID of an IAM user.

aruAssumedRoleId :: Lens' AssumedRoleUser Text

aruArn :: Lens' AssumedRoleUser Text

cSessionToken :: Lens' Credentials Text

cSecretAccessKey :: Lens' Credentials Text

cExpiration :: Lens' Credentials ISO8601

cAccessKeyId :: Lens' Credentials Text

fuFederatedUserId :: Lens' FederatedUser Text

fuArn :: Lens' FederatedUser Text