Network.AWS.SWF.CountOpenWorkflowExecutions

Description

Returns the number of open workflow executions within the given domain that meet the specified filtering criteria.

This operation is eventually consistent. The results are best effort and may not exactly reflect recent updates and changes.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

tagFilter.tag: String constraint. The key is swf:tagFilter.tag.

typeFilter.name: String constraint. The key is swf:typeFilter.name.

typeFilter.version: String constraint. The key is swf:typeFilter.version.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

Creating a Request

data CountOpenWorkflowExecutions

See: newCountOpenWorkflowExecutions smart constructor.

Constructors

CountOpenWorkflowExecutions'

Fields

typeFilter :: Maybe WorkflowTypeFilter

Specifies the type of the workflow executions to be counted.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

tagFilter :: Maybe TagFilter

If specified, only executions that have a tag that matches the filter are counted.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

executionFilter :: Maybe WorkflowExecutionFilter

If specified, only workflow executions matching the WorkflowId in the filter are counted.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

domain :: Text

The name of the domain containing the workflow executions to count.

startTimeFilter :: ExecutionTimeFilter

Specifies the start time criteria that workflow executions must meet in order to be counted.

newCountOpenWorkflowExecutions

Arguments

:: Text
$sel:domain:CountOpenWorkflowExecutions'

-> ExecutionTimeFilter
$sel:startTimeFilter:CountOpenWorkflowExecutions'

-> CountOpenWorkflowExecutions

Create a value of CountOpenWorkflowExecutions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeFilter:CountOpenWorkflowExecutions', countOpenWorkflowExecutions_typeFilter - Specifies the type of the workflow executions to be counted.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

$sel:tagFilter:CountOpenWorkflowExecutions', countOpenWorkflowExecutions_tagFilter - If specified, only executions that have a tag that matches the filter are counted.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

$sel:executionFilter:CountOpenWorkflowExecutions', countOpenWorkflowExecutions_executionFilter - If specified, only workflow executions matching the WorkflowId in the filter are counted.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

$sel:domain:CountOpenWorkflowExecutions', countOpenWorkflowExecutions_domain - The name of the domain containing the workflow executions to count.

$sel:startTimeFilter:CountOpenWorkflowExecutions', countOpenWorkflowExecutions_startTimeFilter - Specifies the start time criteria that workflow executions must meet in order to be counted.

Request Lenses

countOpenWorkflowExecutions_typeFilter :: Lens' CountOpenWorkflowExecutions (Maybe WorkflowTypeFilter)

Specifies the type of the workflow executions to be counted.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

countOpenWorkflowExecutions_tagFilter :: Lens' CountOpenWorkflowExecutions (Maybe TagFilter)

If specified, only executions that have a tag that matches the filter are counted.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

countOpenWorkflowExecutions_executionFilter :: Lens' CountOpenWorkflowExecutions (Maybe WorkflowExecutionFilter)

If specified, only workflow executions matching the WorkflowId in the filter are counted.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

countOpenWorkflowExecutions_domain :: Lens' CountOpenWorkflowExecutions Text

The name of the domain containing the workflow executions to count.

countOpenWorkflowExecutions_startTimeFilter :: Lens' CountOpenWorkflowExecutions ExecutionTimeFilter

Specifies the start time criteria that workflow executions must meet in order to be counted.

Destructuring the Response

data WorkflowExecutionCount

Contains the count of workflow executions returned from CountOpenWorkflowExecutions or CountClosedWorkflowExecutions

See: newWorkflowExecutionCount smart constructor.

Constructors

WorkflowExecutionCount'

Fields

truncated :: Maybe Bool

If set to true, indicates that the actual count was more than the maximum supported by this API and the count returned is the truncated value.

count :: Natural

The number of workflow executions.

newWorkflowExecutionCount

Arguments

:: Natural
$sel:count:WorkflowExecutionCount'

-> WorkflowExecutionCount

Create a value of WorkflowExecutionCount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:truncated:WorkflowExecutionCount', workflowExecutionCount_truncated - If set to true, indicates that the actual count was more than the maximum supported by this API and the count returned is the truncated value.

$sel:count:WorkflowExecutionCount', workflowExecutionCount_count - The number of workflow executions.

Response Lenses

workflowExecutionCount_truncated :: Lens' WorkflowExecutionCount (Maybe Bool)

If set to true, indicates that the actual count was more than the maximum supported by this API and the count returned is the truncated value.

workflowExecutionCount_count :: Lens' WorkflowExecutionCount Natural

The number of workflow executions.