Network.AWS.SWF.DescribeWorkflowExecution

Description

Returns information about the specified workflow execution including its type and some statistics.

This operation is eventually consistent. The results are best effort and may not exactly reflect recent updates and changes.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

Creating a Request

data DescribeWorkflowExecution

See: newDescribeWorkflowExecution smart constructor.

Constructors

DescribeWorkflowExecution'

Fields

domain :: Text

The name of the domain containing the workflow execution.

execution :: WorkflowExecution

The workflow execution to describe.

newDescribeWorkflowExecution

Arguments

:: Text
$sel:domain:DescribeWorkflowExecution'

-> WorkflowExecution
$sel:execution:DescribeWorkflowExecution'

-> DescribeWorkflowExecution

Create a value of DescribeWorkflowExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domain:DescribeWorkflowExecution', describeWorkflowExecution_domain - The name of the domain containing the workflow execution.

$sel:execution:DescribeWorkflowExecution', describeWorkflowExecution_execution - The workflow execution to describe.

Request Lenses

describeWorkflowExecution_domain :: Lens' DescribeWorkflowExecution Text

The name of the domain containing the workflow execution.

describeWorkflowExecution_execution :: Lens' DescribeWorkflowExecution WorkflowExecution

The workflow execution to describe.

Destructuring the Response

data DescribeWorkflowExecutionResponse

Contains details about a workflow execution.

See: newDescribeWorkflowExecutionResponse smart constructor.

Constructors

DescribeWorkflowExecutionResponse'

Fields

latestExecutionContext :: Maybe Text

The latest executionContext provided by the decider for this workflow execution. A decider can provide an executionContext (a free-form string) when closing a decision task using RespondDecisionTaskCompleted.

latestActivityTaskTimestamp :: Maybe POSIX

The time when the last activity task was scheduled for this workflow execution. You can use this information to determine if the workflow has not made progress for an unusually long period of time and might require a corrective action.

httpStatus :: Int

The response's http status code.

executionInfo :: WorkflowExecutionInfo

Information about the workflow execution.

executionConfiguration :: WorkflowExecutionConfiguration

The configuration settings for this workflow execution including timeout values, tasklist etc.

openCounts :: WorkflowExecutionOpenCounts

The number of tasks for this workflow execution. This includes open and closed tasks of all types.

newDescribeWorkflowExecutionResponse

Arguments

:: Int
$sel:httpStatus:DescribeWorkflowExecutionResponse'

-> WorkflowExecutionInfo
$sel:executionInfo:DescribeWorkflowExecutionResponse'

-> WorkflowExecutionConfiguration
$sel:executionConfiguration:DescribeWorkflowExecutionResponse'

-> WorkflowExecutionOpenCounts
$sel:openCounts:DescribeWorkflowExecutionResponse'

-> DescribeWorkflowExecutionResponse

Create a value of DescribeWorkflowExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:latestExecutionContext:DescribeWorkflowExecutionResponse', describeWorkflowExecutionResponse_latestExecutionContext - The latest executionContext provided by the decider for this workflow execution. A decider can provide an executionContext (a free-form string) when closing a decision task using RespondDecisionTaskCompleted.

$sel:latestActivityTaskTimestamp:DescribeWorkflowExecutionResponse', describeWorkflowExecutionResponse_latestActivityTaskTimestamp - The time when the last activity task was scheduled for this workflow execution. You can use this information to determine if the workflow has not made progress for an unusually long period of time and might require a corrective action.

$sel:httpStatus:DescribeWorkflowExecutionResponse', describeWorkflowExecutionResponse_httpStatus - The response's http status code.

$sel:executionInfo:DescribeWorkflowExecutionResponse', describeWorkflowExecutionResponse_executionInfo - Information about the workflow execution.

$sel:executionConfiguration:DescribeWorkflowExecutionResponse', describeWorkflowExecutionResponse_executionConfiguration - The configuration settings for this workflow execution including timeout values, tasklist etc.

$sel:openCounts:DescribeWorkflowExecutionResponse', describeWorkflowExecutionResponse_openCounts - The number of tasks for this workflow execution. This includes open and closed tasks of all types.

Response Lenses

describeWorkflowExecutionResponse_latestExecutionContext :: Lens' DescribeWorkflowExecutionResponse (Maybe Text)

The latest executionContext provided by the decider for this workflow execution. A decider can provide an executionContext (a free-form string) when closing a decision task using RespondDecisionTaskCompleted.

describeWorkflowExecutionResponse_latestActivityTaskTimestamp :: Lens' DescribeWorkflowExecutionResponse (Maybe UTCTime)

The time when the last activity task was scheduled for this workflow execution. You can use this information to determine if the workflow has not made progress for an unusually long period of time and might require a corrective action.

describeWorkflowExecutionResponse_httpStatus :: Lens' DescribeWorkflowExecutionResponse Int

The response's http status code.

describeWorkflowExecutionResponse_executionInfo :: Lens' DescribeWorkflowExecutionResponse WorkflowExecutionInfo

Information about the workflow execution.

describeWorkflowExecutionResponse_executionConfiguration :: Lens' DescribeWorkflowExecutionResponse WorkflowExecutionConfiguration

The configuration settings for this workflow execution including timeout values, tasklist etc.

describeWorkflowExecutionResponse_openCounts :: Lens' DescribeWorkflowExecutionResponse WorkflowExecutionOpenCounts

The number of tasks for this workflow execution. This includes open and closed tasks of all types.