Network.AWS.SWF.GetWorkflowExecutionHistory

Description

Returns the history of the specified workflow execution. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextPageToken returned by the initial call.

This operation is eventually consistent. The results are best effort and may not exactly reflect recent updates and changes.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

This operation returns paginated results.

Creating a Request

data GetWorkflowExecutionHistory

See: newGetWorkflowExecutionHistory smart constructor.

Constructors

GetWorkflowExecutionHistory'

Fields

nextPageToken :: Maybe Text

If NextPageToken is returned there are more results available. The value of NextPageToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 60 seconds. Using an expired pagination token will return a 400 error: "Specified token has exceeded its maximum lifetime".

The configured maximumPageSize determines how many results can be returned in a single call.

maximumPageSize :: Maybe Natural

The maximum number of results that are returned per call. Use nextPageToken to obtain further pages of results.

reverseOrder :: Maybe Bool

When set to true, returns the events in reverse order. By default the results are returned in ascending order of the eventTimeStamp of the events.

domain :: Text

The name of the domain containing the workflow execution.

execution :: WorkflowExecution

Specifies the workflow execution for which to return the history.

newGetWorkflowExecutionHistory

Arguments

:: Text
$sel:domain:GetWorkflowExecutionHistory'

-> WorkflowExecution
$sel:execution:GetWorkflowExecutionHistory'

-> GetWorkflowExecutionHistory

Create a value of GetWorkflowExecutionHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:GetWorkflowExecutionHistory', getWorkflowExecutionHistory_nextPageToken - If NextPageToken is returned there are more results available. The value of NextPageToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 60 seconds. Using an expired pagination token will return a 400 error: "Specified token has exceeded its maximum lifetime".

The configured maximumPageSize determines how many results can be returned in a single call.

$sel:maximumPageSize:GetWorkflowExecutionHistory', getWorkflowExecutionHistory_maximumPageSize - The maximum number of results that are returned per call. Use nextPageToken to obtain further pages of results.

$sel:reverseOrder:GetWorkflowExecutionHistory', getWorkflowExecutionHistory_reverseOrder - When set to true, returns the events in reverse order. By default the results are returned in ascending order of the eventTimeStamp of the events.

$sel:domain:GetWorkflowExecutionHistory', getWorkflowExecutionHistory_domain - The name of the domain containing the workflow execution.

$sel:execution:GetWorkflowExecutionHistory', getWorkflowExecutionHistory_execution - Specifies the workflow execution for which to return the history.

Request Lenses

getWorkflowExecutionHistory_nextPageToken :: Lens' GetWorkflowExecutionHistory (Maybe Text)

If NextPageToken is returned there are more results available. The value of NextPageToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 60 seconds. Using an expired pagination token will return a 400 error: "Specified token has exceeded its maximum lifetime".

The configured maximumPageSize determines how many results can be returned in a single call.

getWorkflowExecutionHistory_maximumPageSize :: Lens' GetWorkflowExecutionHistory (Maybe Natural)

The maximum number of results that are returned per call. Use nextPageToken to obtain further pages of results.

getWorkflowExecutionHistory_reverseOrder :: Lens' GetWorkflowExecutionHistory (Maybe Bool)

When set to true, returns the events in reverse order. By default the results are returned in ascending order of the eventTimeStamp of the events.

getWorkflowExecutionHistory_domain :: Lens' GetWorkflowExecutionHistory Text

The name of the domain containing the workflow execution.

getWorkflowExecutionHistory_execution :: Lens' GetWorkflowExecutionHistory WorkflowExecution

Specifies the workflow execution for which to return the history.

Destructuring the Response

data GetWorkflowExecutionHistoryResponse

Paginated representation of a workflow history for a workflow execution. This is the up to date, complete and authoritative record of the events related to all tasks and events in the life of the workflow execution.

See: newGetWorkflowExecutionHistoryResponse smart constructor.

Constructors

GetWorkflowExecutionHistoryResponse'

Fields

nextPageToken :: Maybe Text

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

httpStatus :: Int

The response's http status code.

events :: [HistoryEvent]

The list of history events.

newGetWorkflowExecutionHistoryResponse

Arguments

:: Int
$sel:httpStatus:GetWorkflowExecutionHistoryResponse'

-> GetWorkflowExecutionHistoryResponse

Create a value of GetWorkflowExecutionHistoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:GetWorkflowExecutionHistory', getWorkflowExecutionHistoryResponse_nextPageToken - If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

$sel:httpStatus:GetWorkflowExecutionHistoryResponse', getWorkflowExecutionHistoryResponse_httpStatus - The response's http status code.

$sel:events:GetWorkflowExecutionHistoryResponse', getWorkflowExecutionHistoryResponse_events - The list of history events.

Response Lenses

getWorkflowExecutionHistoryResponse_nextPageToken :: Lens' GetWorkflowExecutionHistoryResponse (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

getWorkflowExecutionHistoryResponse_httpStatus :: Lens' GetWorkflowExecutionHistoryResponse Int

The response's http status code.

getWorkflowExecutionHistoryResponse_events :: Lens' GetWorkflowExecutionHistoryResponse [HistoryEvent]

The list of history events.