Network.AWS.SWF.ListOpenWorkflowExecutions

Description

Returns a list of open workflow executions in the specified domain that meet the filtering criteria. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextPageToken returned by the initial call.

This operation is eventually consistent. The results are best effort and may not exactly reflect recent updates and changes.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

tagFilter.tag: String constraint. The key is swf:tagFilter.tag.

typeFilter.name: String constraint. The key is swf:typeFilter.name.

typeFilter.version: String constraint. The key is swf:typeFilter.version.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

This operation returns paginated results.

Creating a Request

data ListOpenWorkflowExecutions

See: newListOpenWorkflowExecutions smart constructor.

Constructors

ListOpenWorkflowExecutions'

Fields

typeFilter :: Maybe WorkflowTypeFilter

If specified, only executions of the type specified in the filter are returned.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

tagFilter :: Maybe TagFilter

If specified, only executions that have the matching tag are listed.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

nextPageToken :: Maybe Text

If NextPageToken is returned there are more results available. The value of NextPageToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 60 seconds. Using an expired pagination token will return a 400 error: "Specified token has exceeded its maximum lifetime".

The configured maximumPageSize determines how many results can be returned in a single call.

maximumPageSize :: Maybe Natural

The maximum number of results that are returned per call. Use nextPageToken to obtain further pages of results.

reverseOrder :: Maybe Bool

When set to true, returns the results in reverse order. By default the results are returned in descending order of the start time of the executions.

executionFilter :: Maybe WorkflowExecutionFilter

If specified, only workflow executions matching the workflow ID specified in the filter are returned.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

domain :: Text

The name of the domain that contains the workflow executions to list.

startTimeFilter :: ExecutionTimeFilter

Workflow executions are included in the returned results based on whether their start times are within the range specified by this filter.

newListOpenWorkflowExecutions

Arguments

:: Text
$sel:domain:ListOpenWorkflowExecutions'

-> ExecutionTimeFilter
$sel:startTimeFilter:ListOpenWorkflowExecutions'

-> ListOpenWorkflowExecutions

Create a value of ListOpenWorkflowExecutions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeFilter:ListOpenWorkflowExecutions', listOpenWorkflowExecutions_typeFilter - If specified, only executions of the type specified in the filter are returned.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

$sel:tagFilter:ListOpenWorkflowExecutions', listOpenWorkflowExecutions_tagFilter - If specified, only executions that have the matching tag are listed.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

$sel:nextPageToken:ListOpenWorkflowExecutions', listOpenWorkflowExecutions_nextPageToken - If NextPageToken is returned there are more results available. The value of NextPageToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 60 seconds. Using an expired pagination token will return a 400 error: "Specified token has exceeded its maximum lifetime".

The configured maximumPageSize determines how many results can be returned in a single call.

$sel:maximumPageSize:ListOpenWorkflowExecutions', listOpenWorkflowExecutions_maximumPageSize - The maximum number of results that are returned per call. Use nextPageToken to obtain further pages of results.

$sel:reverseOrder:ListOpenWorkflowExecutions', listOpenWorkflowExecutions_reverseOrder - When set to true, returns the results in reverse order. By default the results are returned in descending order of the start time of the executions.

$sel:executionFilter:ListOpenWorkflowExecutions', listOpenWorkflowExecutions_executionFilter - If specified, only workflow executions matching the workflow ID specified in the filter are returned.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

$sel:domain:ListOpenWorkflowExecutions', listOpenWorkflowExecutions_domain - The name of the domain that contains the workflow executions to list.

$sel:startTimeFilter:ListOpenWorkflowExecutions', listOpenWorkflowExecutions_startTimeFilter - Workflow executions are included in the returned results based on whether their start times are within the range specified by this filter.

Request Lenses

listOpenWorkflowExecutions_typeFilter :: Lens' ListOpenWorkflowExecutions (Maybe WorkflowTypeFilter)

If specified, only executions of the type specified in the filter are returned.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

listOpenWorkflowExecutions_tagFilter :: Lens' ListOpenWorkflowExecutions (Maybe TagFilter)

If specified, only executions that have the matching tag are listed.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

listOpenWorkflowExecutions_nextPageToken :: Lens' ListOpenWorkflowExecutions (Maybe Text)

If NextPageToken is returned there are more results available. The value of NextPageToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 60 seconds. Using an expired pagination token will return a 400 error: "Specified token has exceeded its maximum lifetime".

The configured maximumPageSize determines how many results can be returned in a single call.

listOpenWorkflowExecutions_maximumPageSize :: Lens' ListOpenWorkflowExecutions (Maybe Natural)

The maximum number of results that are returned per call. Use nextPageToken to obtain further pages of results.

listOpenWorkflowExecutions_reverseOrder :: Lens' ListOpenWorkflowExecutions (Maybe Bool)

When set to true, returns the results in reverse order. By default the results are returned in descending order of the start time of the executions.

listOpenWorkflowExecutions_executionFilter :: Lens' ListOpenWorkflowExecutions (Maybe WorkflowExecutionFilter)

If specified, only workflow executions matching the workflow ID specified in the filter are returned.

executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

listOpenWorkflowExecutions_domain :: Lens' ListOpenWorkflowExecutions Text

The name of the domain that contains the workflow executions to list.

listOpenWorkflowExecutions_startTimeFilter :: Lens' ListOpenWorkflowExecutions ExecutionTimeFilter

Workflow executions are included in the returned results based on whether their start times are within the range specified by this filter.

Destructuring the Response

data WorkflowExecutionInfos

Contains a paginated list of information about workflow executions.

See: newWorkflowExecutionInfos smart constructor.

Constructors

WorkflowExecutionInfos'

Fields

nextPageToken :: Maybe Text

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

executionInfos :: [WorkflowExecutionInfo]

The list of workflow information structures.

newWorkflowExecutionInfos :: WorkflowExecutionInfos

Create a value of WorkflowExecutionInfos with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:WorkflowExecutionInfos', workflowExecutionInfos_nextPageToken - If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

$sel:executionInfos:WorkflowExecutionInfos', workflowExecutionInfos_executionInfos - The list of workflow information structures.

Response Lenses

workflowExecutionInfos_nextPageToken :: Lens' WorkflowExecutionInfos (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

workflowExecutionInfos_executionInfos :: Lens' WorkflowExecutionInfos [WorkflowExecutionInfo]

The list of workflow information structures.