Network.AWS.SWF.PollForActivityTask

Description

Used by workers to get an ActivityTask from the specified activity taskList. This initiates a long poll, where the service holds the HTTP connection open and responds as soon as a task becomes available. The maximum time the service holds on to the request before responding is 60 seconds. If no task is available within 60 seconds, the poll returns an empty result. An empty result, in this context, means that an ActivityTask is returned, but that the value of taskToken is an empty string. If a task is returned, the worker should use its type to identify and process it correctly.

Workers should set their client side socket timeout to at least 70 seconds (10 seconds higher than the maximum time service may hold the poll request).

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the taskList.name parameter by using a Condition element with the swf:taskList.name key to allow the action to access only certain task lists.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

Creating a Request

data PollForActivityTask

See: newPollForActivityTask smart constructor.

Constructors

PollForActivityTask'

Fields

identity :: Maybe Text

Identity of the worker making the request, recorded in the ActivityTaskStarted event in the workflow history. This enables diagnostic tracing when problems arise. The form of this identity is user defined.

domain :: Text

The name of the domain that contains the task lists being polled.

taskList :: TaskList

Specifies the task list to poll for activity tasks.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not be the literal string arn.

newPollForActivityTask

Arguments

:: Text
$sel:domain:PollForActivityTask'

-> TaskList
$sel:taskList:PollForActivityTask'

-> PollForActivityTask

Create a value of PollForActivityTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identity:PollForActivityTask', pollForActivityTask_identity - Identity of the worker making the request, recorded in the ActivityTaskStarted event in the workflow history. This enables diagnostic tracing when problems arise. The form of this identity is user defined.

$sel:domain:PollForActivityTask', pollForActivityTask_domain - The name of the domain that contains the task lists being polled.

$sel:taskList:PollForActivityTask', pollForActivityTask_taskList - Specifies the task list to poll for activity tasks.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not be the literal string arn.

Request Lenses

pollForActivityTask_identity :: Lens' PollForActivityTask (Maybe Text)

Identity of the worker making the request, recorded in the ActivityTaskStarted event in the workflow history. This enables diagnostic tracing when problems arise. The form of this identity is user defined.

pollForActivityTask_domain :: Lens' PollForActivityTask Text

The name of the domain that contains the task lists being polled.

pollForActivityTask_taskList :: Lens' PollForActivityTask TaskList

Specifies the task list to poll for activity tasks.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not be the literal string arn.

Destructuring the Response

data PollForActivityTaskResponse

Unit of work sent to an activity worker.

See: newPollForActivityTaskResponse smart constructor.

Constructors

PollForActivityTaskResponse'

Fields

activityType :: Maybe ActivityType

The type of this activity task.

activityId :: Maybe Text

The unique ID of the task.

input :: Maybe Text

The inputs provided when the activity task was scheduled. The form of the input is user defined and should be meaningful to the activity implementation.

taskToken :: Maybe Text

The opaque string used as a handle on the task. This token is used by workers to communicate progress and response information back to the system about the task.

workflowExecution :: Maybe WorkflowExecution

The workflow execution that started this activity task.

httpStatus :: Int

The response's http status code.

startedEventId :: Integer

The ID of the ActivityTaskStarted event recorded in the history.

newPollForActivityTaskResponse

Arguments

:: Int
$sel:httpStatus:PollForActivityTaskResponse'

-> Integer
$sel:startedEventId:PollForActivityTaskResponse'

-> PollForActivityTaskResponse

Create a value of PollForActivityTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:activityType:PollForActivityTaskResponse', pollForActivityTaskResponse_activityType - The type of this activity task.

$sel:activityId:PollForActivityTaskResponse', pollForActivityTaskResponse_activityId - The unique ID of the task.

$sel:input:PollForActivityTaskResponse', pollForActivityTaskResponse_input - The inputs provided when the activity task was scheduled. The form of the input is user defined and should be meaningful to the activity implementation.

$sel:taskToken:PollForActivityTaskResponse', pollForActivityTaskResponse_taskToken - The opaque string used as a handle on the task. This token is used by workers to communicate progress and response information back to the system about the task.

$sel:workflowExecution:PollForActivityTaskResponse', pollForActivityTaskResponse_workflowExecution - The workflow execution that started this activity task.

$sel:httpStatus:PollForActivityTaskResponse', pollForActivityTaskResponse_httpStatus - The response's http status code.

$sel:startedEventId:PollForActivityTaskResponse', pollForActivityTaskResponse_startedEventId - The ID of the ActivityTaskStarted event recorded in the history.

Response Lenses

pollForActivityTaskResponse_activityType :: Lens' PollForActivityTaskResponse (Maybe ActivityType)

The type of this activity task.

pollForActivityTaskResponse_activityId :: Lens' PollForActivityTaskResponse (Maybe Text)

The unique ID of the task.

pollForActivityTaskResponse_input :: Lens' PollForActivityTaskResponse (Maybe Text)

The inputs provided when the activity task was scheduled. The form of the input is user defined and should be meaningful to the activity implementation.

pollForActivityTaskResponse_taskToken :: Lens' PollForActivityTaskResponse (Maybe Text)

The opaque string used as a handle on the task. This token is used by workers to communicate progress and response information back to the system about the task.

pollForActivityTaskResponse_workflowExecution :: Lens' PollForActivityTaskResponse (Maybe WorkflowExecution)

The workflow execution that started this activity task.

pollForActivityTaskResponse_httpStatus :: Lens' PollForActivityTaskResponse Int

The response's http status code.

pollForActivityTaskResponse_startedEventId :: Lens' PollForActivityTaskResponse Integer

The ID of the ActivityTaskStarted event recorded in the history.