Network.AWS.SWF.PollForDecisionTask

Description

Used by deciders to get a DecisionTask from the specified decision taskList. A decision task may be returned for any open workflow execution that is using the specified task list. The task includes a paginated view of the history of the workflow execution. The decider should use the workflow type and the history to determine how to properly handle the task.

This action initiates a long poll, where the service holds the HTTP connection open and responds as soon a task becomes available. If no decision task is available in the specified task list before the timeout of 60 seconds expires, an empty result is returned. An empty result, in this context, means that a DecisionTask is returned, but that the value of taskToken is an empty string.

Deciders should set their client side socket timeout to at least 70 seconds (10 seconds higher than the timeout).

Because the number of workflow history events for a single workflow execution might be very large, the result returned might be split up across a number of pages. To retrieve subsequent pages, make additional calls to PollForDecisionTask using the nextPageToken returned by the initial call. Note that you do not call GetWorkflowExecutionHistory with this nextPageToken. Instead, call PollForDecisionTask again.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the taskList.name parameter by using a Condition element with the swf:taskList.name key to allow the action to access only certain task lists.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

This operation returns paginated results.

Creating a Request

data PollForDecisionTask

See: newPollForDecisionTask smart constructor.

Constructors

PollForDecisionTask'

Fields

identity :: Maybe Text

Identity of the decider making the request, which is recorded in the DecisionTaskStarted event in the workflow history. This enables diagnostic tracing when problems arise. The form of this identity is user defined.

nextPageToken :: Maybe Text

If NextPageToken is returned there are more results available. The value of NextPageToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 60 seconds. Using an expired pagination token will return a 400 error: "Specified token has exceeded its maximum lifetime".

The configured maximumPageSize determines how many results can be returned in a single call.

The nextPageToken returned by this action cannot be used with GetWorkflowExecutionHistory to get the next page. You must call PollForDecisionTask again (with the nextPageToken) to retrieve the next page of history records. Calling PollForDecisionTask with a nextPageToken doesn't return a new decision task.

maximumPageSize :: Maybe Natural

The maximum number of results that are returned per call. Use nextPageToken to obtain further pages of results.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

reverseOrder :: Maybe Bool

When set to true, returns the events in reverse order. By default the results are returned in ascending order of the eventTimestamp of the events.

domain :: Text

The name of the domain containing the task lists to poll.

taskList :: TaskList

Specifies the task list to poll for decision tasks.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not be the literal string arn.

newPollForDecisionTask

Arguments

:: Text
$sel:domain:PollForDecisionTask'

-> TaskList
$sel:taskList:PollForDecisionTask'

-> PollForDecisionTask

Create a value of PollForDecisionTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identity:PollForDecisionTask', pollForDecisionTask_identity - Identity of the decider making the request, which is recorded in the DecisionTaskStarted event in the workflow history. This enables diagnostic tracing when problems arise. The form of this identity is user defined.

$sel:nextPageToken:PollForDecisionTask', pollForDecisionTask_nextPageToken - If NextPageToken is returned there are more results available. The value of NextPageToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 60 seconds. Using an expired pagination token will return a 400 error: "Specified token has exceeded its maximum lifetime".

The configured maximumPageSize determines how many results can be returned in a single call.

The nextPageToken returned by this action cannot be used with GetWorkflowExecutionHistory to get the next page. You must call PollForDecisionTask again (with the nextPageToken) to retrieve the next page of history records. Calling PollForDecisionTask with a nextPageToken doesn't return a new decision task.

$sel:maximumPageSize:PollForDecisionTask', pollForDecisionTask_maximumPageSize - The maximum number of results that are returned per call. Use nextPageToken to obtain further pages of results.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

$sel:reverseOrder:PollForDecisionTask', pollForDecisionTask_reverseOrder - When set to true, returns the events in reverse order. By default the results are returned in ascending order of the eventTimestamp of the events.

$sel:domain:PollForDecisionTask', pollForDecisionTask_domain - The name of the domain containing the task lists to poll.

$sel:taskList:PollForDecisionTask', pollForDecisionTask_taskList - Specifies the task list to poll for decision tasks.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not be the literal string arn.

Request Lenses

pollForDecisionTask_identity :: Lens' PollForDecisionTask (Maybe Text)

Identity of the decider making the request, which is recorded in the DecisionTaskStarted event in the workflow history. This enables diagnostic tracing when problems arise. The form of this identity is user defined.

pollForDecisionTask_nextPageToken :: Lens' PollForDecisionTask (Maybe Text)

If NextPageToken is returned there are more results available. The value of NextPageToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 60 seconds. Using an expired pagination token will return a 400 error: "Specified token has exceeded its maximum lifetime".

The configured maximumPageSize determines how many results can be returned in a single call.

The nextPageToken returned by this action cannot be used with GetWorkflowExecutionHistory to get the next page. You must call PollForDecisionTask again (with the nextPageToken) to retrieve the next page of history records. Calling PollForDecisionTask with a nextPageToken doesn't return a new decision task.

pollForDecisionTask_maximumPageSize :: Lens' PollForDecisionTask (Maybe Natural)

The maximum number of results that are returned per call. Use nextPageToken to obtain further pages of results.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

pollForDecisionTask_reverseOrder :: Lens' PollForDecisionTask (Maybe Bool)

When set to true, returns the events in reverse order. By default the results are returned in ascending order of the eventTimestamp of the events.

pollForDecisionTask_domain :: Lens' PollForDecisionTask Text

The name of the domain containing the task lists to poll.

pollForDecisionTask_taskList :: Lens' PollForDecisionTask TaskList

Specifies the task list to poll for decision tasks.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not be the literal string arn.

Destructuring the Response

data PollForDecisionTaskResponse

A structure that represents a decision task. Decision tasks are sent to deciders in order for them to make decisions.

See: newPollForDecisionTaskResponse smart constructor.

Constructors

PollForDecisionTaskResponse'

Fields

previousStartedEventId :: Maybe Integer

The ID of the DecisionTaskStarted event of the previous decision task of this workflow execution that was processed by the decider. This can be used to determine the events in the history new since the last decision task received by the decider.

workflowExecution :: Maybe WorkflowExecution

The workflow execution for which this decision task was created.

workflowType :: Maybe WorkflowType

The type of the workflow execution for which this decision task was created.

nextPageToken :: Maybe Text

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

events :: Maybe [HistoryEvent]

A paginated list of history events of the workflow execution. The decider uses this during the processing of the decision task.

taskToken :: Maybe Text

The opaque string used as a handle on the task. This token is used by workers to communicate progress and response information back to the system about the task.

httpStatus :: Int

The response's http status code.

startedEventId :: Integer

The ID of the DecisionTaskStarted event recorded in the history.

newPollForDecisionTaskResponse

Arguments

:: Int
$sel:httpStatus:PollForDecisionTaskResponse'

-> Integer
$sel:startedEventId:PollForDecisionTaskResponse'

-> PollForDecisionTaskResponse

Create a value of PollForDecisionTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:previousStartedEventId:PollForDecisionTaskResponse', pollForDecisionTaskResponse_previousStartedEventId - The ID of the DecisionTaskStarted event of the previous decision task of this workflow execution that was processed by the decider. This can be used to determine the events in the history new since the last decision task received by the decider.

$sel:workflowExecution:PollForDecisionTaskResponse', pollForDecisionTaskResponse_workflowExecution - The workflow execution for which this decision task was created.

$sel:workflowType:PollForDecisionTaskResponse', pollForDecisionTaskResponse_workflowType - The type of the workflow execution for which this decision task was created.

$sel:nextPageToken:PollForDecisionTask', pollForDecisionTaskResponse_nextPageToken - If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

$sel:events:PollForDecisionTaskResponse', pollForDecisionTaskResponse_events - A paginated list of history events of the workflow execution. The decider uses this during the processing of the decision task.

$sel:taskToken:PollForDecisionTaskResponse', pollForDecisionTaskResponse_taskToken - The opaque string used as a handle on the task. This token is used by workers to communicate progress and response information back to the system about the task.

$sel:httpStatus:PollForDecisionTaskResponse', pollForDecisionTaskResponse_httpStatus - The response's http status code.

$sel:startedEventId:PollForDecisionTaskResponse', pollForDecisionTaskResponse_startedEventId - The ID of the DecisionTaskStarted event recorded in the history.

Response Lenses

pollForDecisionTaskResponse_previousStartedEventId :: Lens' PollForDecisionTaskResponse (Maybe Integer)

The ID of the DecisionTaskStarted event of the previous decision task of this workflow execution that was processed by the decider. This can be used to determine the events in the history new since the last decision task received by the decider.

pollForDecisionTaskResponse_workflowExecution :: Lens' PollForDecisionTaskResponse (Maybe WorkflowExecution)

The workflow execution for which this decision task was created.

pollForDecisionTaskResponse_workflowType :: Lens' PollForDecisionTaskResponse (Maybe WorkflowType)

The type of the workflow execution for which this decision task was created.

pollForDecisionTaskResponse_nextPageToken :: Lens' PollForDecisionTaskResponse (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

pollForDecisionTaskResponse_events :: Lens' PollForDecisionTaskResponse (Maybe [HistoryEvent])

A paginated list of history events of the workflow execution. The decider uses this during the processing of the decision task.

pollForDecisionTaskResponse_taskToken :: Lens' PollForDecisionTaskResponse (Maybe Text)

The opaque string used as a handle on the task. This token is used by workers to communicate progress and response information back to the system about the task.

pollForDecisionTaskResponse_httpStatus :: Lens' PollForDecisionTaskResponse Int

The response's http status code.

pollForDecisionTaskResponse_startedEventId :: Lens' PollForDecisionTaskResponse Integer

The ID of the DecisionTaskStarted event recorded in the history.