Network.AWS.SWF.RegisterDomain

Description

Registers a new domain.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

You cannot use an IAM policy to control domain access for this action. The name of the domain being registered is available as the resource of this action.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

Creating a Request

data RegisterDomain

See: newRegisterDomain smart constructor.

Constructors

RegisterDomain'

Fields

tags :: Maybe [ResourceTag]

Tags to be added when registering a domain.

Tags may only contain unicode letters, digits, whitespace, or these symbols: _ . : / = + - @.

description :: Maybe Text

A text description of the domain.

name :: Text

Name of the domain to register. The name must be unique in the region that the domain is registered in.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not be the literal string arn.

workflowExecutionRetentionPeriodInDays :: Text

The duration (in days) that records and histories of workflow executions on the domain should be kept by the service. After the retention period, the workflow execution isn't available in the results of visibility calls.

If you pass the value NONE or 0 (zero), then the workflow execution history isn't retained. As soon as the workflow execution completes, the execution record and its history are deleted.

The maximum workflow execution retention period is 90 days. For more information about Amazon SWF service limits, see: Amazon SWF Service Limits in the Amazon SWF Developer Guide.

newRegisterDomain

Arguments

:: Text
$sel:name:RegisterDomain'

-> Text
$sel:workflowExecutionRetentionPeriodInDays:RegisterDomain'

-> RegisterDomain

Create a value of RegisterDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:RegisterDomain', registerDomain_tags - Tags to be added when registering a domain.

Tags may only contain unicode letters, digits, whitespace, or these symbols: _ . : / = + - @.

$sel:description:RegisterDomain', registerDomain_description - A text description of the domain.

$sel:name:RegisterDomain', registerDomain_name - Name of the domain to register. The name must be unique in the region that the domain is registered in.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not be the literal string arn.

$sel:workflowExecutionRetentionPeriodInDays:RegisterDomain', registerDomain_workflowExecutionRetentionPeriodInDays - The duration (in days) that records and histories of workflow executions on the domain should be kept by the service. After the retention period, the workflow execution isn't available in the results of visibility calls.

If you pass the value NONE or 0 (zero), then the workflow execution history isn't retained. As soon as the workflow execution completes, the execution record and its history are deleted.

The maximum workflow execution retention period is 90 days. For more information about Amazon SWF service limits, see: Amazon SWF Service Limits in the Amazon SWF Developer Guide.

Request Lenses

registerDomain_tags :: Lens' RegisterDomain (Maybe [ResourceTag])

Tags to be added when registering a domain.

Tags may only contain unicode letters, digits, whitespace, or these symbols: _ . : / = + - @.

registerDomain_description :: Lens' RegisterDomain (Maybe Text)

A text description of the domain.

registerDomain_name :: Lens' RegisterDomain Text

Name of the domain to register. The name must be unique in the region that the domain is registered in.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not be the literal string arn.

registerDomain_workflowExecutionRetentionPeriodInDays :: Lens' RegisterDomain Text

The duration (in days) that records and histories of workflow executions on the domain should be kept by the service. After the retention period, the workflow execution isn't available in the results of visibility calls.

If you pass the value NONE or 0 (zero), then the workflow execution history isn't retained. As soon as the workflow execution completes, the execution record and its history are deleted.

The maximum workflow execution retention period is 90 days. For more information about Amazon SWF service limits, see: Amazon SWF Service Limits in the Amazon SWF Developer Guide.

Destructuring the Response

data RegisterDomainResponse

See: newRegisterDomainResponse smart constructor.

Constructors

RegisterDomainResponse'

newRegisterDomainResponse :: RegisterDomainResponse

Create a value of RegisterDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.