Network.AWS.SWF.Types.ActivityTaskFailedEventAttributes

Description

Documentation

data ActivityTaskFailedEventAttributes

Provides the details of the ActivityTaskFailed event.

See: newActivityTaskFailedEventAttributes smart constructor.

Constructors

ActivityTaskFailedEventAttributes'

Fields

details :: Maybe Text

The details of the failure.

reason :: Maybe Text

The reason provided for the failure.

scheduledEventId :: Integer

The ID of the ActivityTaskScheduled event that was recorded when this activity task was scheduled. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

startedEventId :: Integer

The ID of the ActivityTaskStarted event recorded when this activity task was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

newActivityTaskFailedEventAttributes

Arguments

:: Integer
$sel:scheduledEventId:ActivityTaskFailedEventAttributes'

-> Integer
$sel:startedEventId:ActivityTaskFailedEventAttributes'

-> ActivityTaskFailedEventAttributes

Create a value of ActivityTaskFailedEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:details:ActivityTaskFailedEventAttributes', activityTaskFailedEventAttributes_details - The details of the failure.

$sel:reason:ActivityTaskFailedEventAttributes', activityTaskFailedEventAttributes_reason - The reason provided for the failure.

$sel:scheduledEventId:ActivityTaskFailedEventAttributes', activityTaskFailedEventAttributes_scheduledEventId - The ID of the ActivityTaskScheduled event that was recorded when this activity task was scheduled. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

$sel:startedEventId:ActivityTaskFailedEventAttributes', activityTaskFailedEventAttributes_startedEventId - The ID of the ActivityTaskStarted event recorded when this activity task was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

activityTaskFailedEventAttributes_details :: Lens' ActivityTaskFailedEventAttributes (Maybe Text)

The details of the failure.

activityTaskFailedEventAttributes_reason :: Lens' ActivityTaskFailedEventAttributes (Maybe Text)

The reason provided for the failure.

activityTaskFailedEventAttributes_scheduledEventId :: Lens' ActivityTaskFailedEventAttributes Integer

The ID of the ActivityTaskScheduled event that was recorded when this activity task was scheduled. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

activityTaskFailedEventAttributes_startedEventId :: Lens' ActivityTaskFailedEventAttributes Integer

The ID of the ActivityTaskStarted event recorded when this activity task was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.