Network.AWS.SWF.Types.ActivityTaskScheduledEventAttributes

Description

Documentation

data ActivityTaskScheduledEventAttributes

Provides the details of the ActivityTaskScheduled event.

See: newActivityTaskScheduledEventAttributes smart constructor.

Constructors

ActivityTaskScheduledEventAttributes'

Fields

input :: Maybe Text

The input provided to the activity task.

heartbeatTimeout :: Maybe Text

The maximum time before which the worker processing this task must report progress by calling RecordActivityTaskHeartbeat. If the timeout is exceeded, the activity task is automatically timed out. If the worker subsequently attempts to record a heartbeat or return a result, it is ignored.

scheduleToCloseTimeout :: Maybe Text

The maximum amount of time for this activity task.

scheduleToStartTimeout :: Maybe Text

The maximum amount of time the activity task can wait to be assigned to a worker.

taskPriority :: Maybe Text

The priority to assign to the scheduled activity task. If set, this overrides any default priority value that was assigned when the activity type was registered.

Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

control :: Maybe Text

Data attached to the event that can be used by the decider in subsequent workflow tasks. This data isn't sent to the activity.

startToCloseTimeout :: Maybe Text

The maximum amount of time a worker may take to process the activity task.

activityType :: ActivityType

The type of the activity task.

activityId :: Text

The unique ID of the activity task.

taskList :: TaskList

The task list in which the activity task has been scheduled.

decisionTaskCompletedEventId :: Integer

The ID of the DecisionTaskCompleted event corresponding to the decision that resulted in the scheduling of this activity task. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

newActivityTaskScheduledEventAttributes

Arguments

:: ActivityType
$sel:activityType:ActivityTaskScheduledEventAttributes'

-> Text
$sel:activityId:ActivityTaskScheduledEventAttributes'

-> TaskList
$sel:taskList:ActivityTaskScheduledEventAttributes'

-> Integer
$sel:decisionTaskCompletedEventId:ActivityTaskScheduledEventAttributes'

-> ActivityTaskScheduledEventAttributes

Create a value of ActivityTaskScheduledEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:input:ActivityTaskScheduledEventAttributes', activityTaskScheduledEventAttributes_input - The input provided to the activity task.

$sel:heartbeatTimeout:ActivityTaskScheduledEventAttributes', activityTaskScheduledEventAttributes_heartbeatTimeout - The maximum time before which the worker processing this task must report progress by calling RecordActivityTaskHeartbeat. If the timeout is exceeded, the activity task is automatically timed out. If the worker subsequently attempts to record a heartbeat or return a result, it is ignored.

$sel:scheduleToCloseTimeout:ActivityTaskScheduledEventAttributes', activityTaskScheduledEventAttributes_scheduleToCloseTimeout - The maximum amount of time for this activity task.

$sel:scheduleToStartTimeout:ActivityTaskScheduledEventAttributes', activityTaskScheduledEventAttributes_scheduleToStartTimeout - The maximum amount of time the activity task can wait to be assigned to a worker.

$sel:taskPriority:ActivityTaskScheduledEventAttributes', activityTaskScheduledEventAttributes_taskPriority - The priority to assign to the scheduled activity task. If set, this overrides any default priority value that was assigned when the activity type was registered.

Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

$sel:control:ActivityTaskScheduledEventAttributes', activityTaskScheduledEventAttributes_control - Data attached to the event that can be used by the decider in subsequent workflow tasks. This data isn't sent to the activity.

$sel:startToCloseTimeout:ActivityTaskScheduledEventAttributes', activityTaskScheduledEventAttributes_startToCloseTimeout - The maximum amount of time a worker may take to process the activity task.

$sel:activityType:ActivityTaskScheduledEventAttributes', activityTaskScheduledEventAttributes_activityType - The type of the activity task.

$sel:activityId:ActivityTaskScheduledEventAttributes', activityTaskScheduledEventAttributes_activityId - The unique ID of the activity task.

$sel:taskList:ActivityTaskScheduledEventAttributes', activityTaskScheduledEventAttributes_taskList - The task list in which the activity task has been scheduled.

$sel:decisionTaskCompletedEventId:ActivityTaskScheduledEventAttributes', activityTaskScheduledEventAttributes_decisionTaskCompletedEventId - The ID of the DecisionTaskCompleted event corresponding to the decision that resulted in the scheduling of this activity task. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

activityTaskScheduledEventAttributes_input :: Lens' ActivityTaskScheduledEventAttributes (Maybe Text)

The input provided to the activity task.

activityTaskScheduledEventAttributes_heartbeatTimeout :: Lens' ActivityTaskScheduledEventAttributes (Maybe Text)

The maximum time before which the worker processing this task must report progress by calling RecordActivityTaskHeartbeat. If the timeout is exceeded, the activity task is automatically timed out. If the worker subsequently attempts to record a heartbeat or return a result, it is ignored.

activityTaskScheduledEventAttributes_scheduleToCloseTimeout :: Lens' ActivityTaskScheduledEventAttributes (Maybe Text)

The maximum amount of time for this activity task.

activityTaskScheduledEventAttributes_scheduleToStartTimeout :: Lens' ActivityTaskScheduledEventAttributes (Maybe Text)

The maximum amount of time the activity task can wait to be assigned to a worker.

activityTaskScheduledEventAttributes_taskPriority :: Lens' ActivityTaskScheduledEventAttributes (Maybe Text)

The priority to assign to the scheduled activity task. If set, this overrides any default priority value that was assigned when the activity type was registered.

Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

activityTaskScheduledEventAttributes_control :: Lens' ActivityTaskScheduledEventAttributes (Maybe Text)

Data attached to the event that can be used by the decider in subsequent workflow tasks. This data isn't sent to the activity.

activityTaskScheduledEventAttributes_startToCloseTimeout :: Lens' ActivityTaskScheduledEventAttributes (Maybe Text)

The maximum amount of time a worker may take to process the activity task.

activityTaskScheduledEventAttributes_activityType :: Lens' ActivityTaskScheduledEventAttributes ActivityType

The type of the activity task.

activityTaskScheduledEventAttributes_activityId :: Lens' ActivityTaskScheduledEventAttributes Text

The unique ID of the activity task.

activityTaskScheduledEventAttributes_taskList :: Lens' ActivityTaskScheduledEventAttributes TaskList

The task list in which the activity task has been scheduled.

activityTaskScheduledEventAttributes_decisionTaskCompletedEventId :: Lens' ActivityTaskScheduledEventAttributes Integer

The ID of the DecisionTaskCompleted event corresponding to the decision that resulted in the scheduling of this activity task. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.