Network.AWS.SWF.Types.ActivityTaskTimedOutEventAttributes

Description

Documentation

data ActivityTaskTimedOutEventAttributes

Provides the details of the ActivityTaskTimedOut event.

See: newActivityTaskTimedOutEventAttributes smart constructor.

Constructors

ActivityTaskTimedOutEventAttributes'

Fields

details :: Maybe Text

Contains the content of the details parameter for the last call made by the activity to RecordActivityTaskHeartbeat.

timeoutType :: ActivityTaskTimeoutType

The type of the timeout that caused this event.

scheduledEventId :: Integer

The ID of the ActivityTaskScheduled event that was recorded when this activity task was scheduled. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

startedEventId :: Integer

The ID of the ActivityTaskStarted event recorded when this activity task was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

newActivityTaskTimedOutEventAttributes

Arguments

:: ActivityTaskTimeoutType
$sel:timeoutType:ActivityTaskTimedOutEventAttributes'

-> Integer
$sel:scheduledEventId:ActivityTaskTimedOutEventAttributes'

-> Integer
$sel:startedEventId:ActivityTaskTimedOutEventAttributes'

-> ActivityTaskTimedOutEventAttributes

Create a value of ActivityTaskTimedOutEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:details:ActivityTaskTimedOutEventAttributes', activityTaskTimedOutEventAttributes_details - Contains the content of the details parameter for the last call made by the activity to RecordActivityTaskHeartbeat.

$sel:timeoutType:ActivityTaskTimedOutEventAttributes', activityTaskTimedOutEventAttributes_timeoutType - The type of the timeout that caused this event.

$sel:scheduledEventId:ActivityTaskTimedOutEventAttributes', activityTaskTimedOutEventAttributes_scheduledEventId - The ID of the ActivityTaskScheduled event that was recorded when this activity task was scheduled. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

$sel:startedEventId:ActivityTaskTimedOutEventAttributes', activityTaskTimedOutEventAttributes_startedEventId - The ID of the ActivityTaskStarted event recorded when this activity task was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

activityTaskTimedOutEventAttributes_details :: Lens' ActivityTaskTimedOutEventAttributes (Maybe Text)

Contains the content of the details parameter for the last call made by the activity to RecordActivityTaskHeartbeat.

activityTaskTimedOutEventAttributes_timeoutType :: Lens' ActivityTaskTimedOutEventAttributes ActivityTaskTimeoutType

The type of the timeout that caused this event.

activityTaskTimedOutEventAttributes_scheduledEventId :: Lens' ActivityTaskTimedOutEventAttributes Integer

The ID of the ActivityTaskScheduled event that was recorded when this activity task was scheduled. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

activityTaskTimedOutEventAttributes_startedEventId :: Lens' ActivityTaskTimedOutEventAttributes Integer

The ID of the ActivityTaskStarted event recorded when this activity task was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.