Network.AWS.SWF.Types.ActivityTypeInfo

Description

Documentation

data ActivityTypeInfo

Detailed information about an activity type.

See: newActivityTypeInfo smart constructor.

Constructors

ActivityTypeInfo'

Fields

deprecationDate :: Maybe POSIX

If DEPRECATED, the date and time DeprecateActivityType was called.

description :: Maybe Text

The description of the activity type provided in RegisterActivityType.

activityType :: ActivityType

The ActivityType type structure representing the activity type.

status :: RegistrationStatus

The current status of the activity type.

creationDate :: POSIX

The date and time this activity type was created through RegisterActivityType.

newActivityTypeInfo

Arguments

:: ActivityType
$sel:activityType:ActivityTypeInfo'

-> RegistrationStatus
$sel:status:ActivityTypeInfo'

-> UTCTime
$sel:creationDate:ActivityTypeInfo'

-> ActivityTypeInfo

Create a value of ActivityTypeInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deprecationDate:ActivityTypeInfo', activityTypeInfo_deprecationDate - If DEPRECATED, the date and time DeprecateActivityType was called.

$sel:description:ActivityTypeInfo', activityTypeInfo_description - The description of the activity type provided in RegisterActivityType.

$sel:activityType:ActivityTypeInfo', activityTypeInfo_activityType - The ActivityType type structure representing the activity type.

$sel:status:ActivityTypeInfo', activityTypeInfo_status - The current status of the activity type.

$sel:creationDate:ActivityTypeInfo', activityTypeInfo_creationDate - The date and time this activity type was created through RegisterActivityType.

activityTypeInfo_deprecationDate :: Lens' ActivityTypeInfo (Maybe UTCTime)

If DEPRECATED, the date and time DeprecateActivityType was called.

activityTypeInfo_description :: Lens' ActivityTypeInfo (Maybe Text)

The description of the activity type provided in RegisterActivityType.

activityTypeInfo_activityType :: Lens' ActivityTypeInfo ActivityType

The ActivityType type structure representing the activity type.

activityTypeInfo_status :: Lens' ActivityTypeInfo RegistrationStatus

The current status of the activity type.

activityTypeInfo_creationDate :: Lens' ActivityTypeInfo UTCTime

The date and time this activity type was created through RegisterActivityType.