Network.AWS.SWF.Types.CancelTimerFailedEventAttributes

Description

Documentation

data CancelTimerFailedEventAttributes

Provides the details of the CancelTimerFailed event.

See: newCancelTimerFailedEventAttributes smart constructor.

Constructors

CancelTimerFailedEventAttributes'

Fields

timerId :: Text

The timerId provided in the CancelTimer decision that failed.

cause :: CancelTimerFailedCause

The cause of the failure. This information is generated by the system and can be useful for diagnostic purposes.

If cause is set to OPERATION_NOT_PERMITTED, the decision failed because it lacked sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

decisionTaskCompletedEventId :: Integer

The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the CancelTimer decision to cancel this timer. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

newCancelTimerFailedEventAttributes

Arguments

:: Text
$sel:timerId:CancelTimerFailedEventAttributes'

-> CancelTimerFailedCause
$sel:cause:CancelTimerFailedEventAttributes'

-> Integer
$sel:decisionTaskCompletedEventId:CancelTimerFailedEventAttributes'

-> CancelTimerFailedEventAttributes

Create a value of CancelTimerFailedEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timerId:CancelTimerFailedEventAttributes', cancelTimerFailedEventAttributes_timerId - The timerId provided in the CancelTimer decision that failed.

$sel:cause:CancelTimerFailedEventAttributes', cancelTimerFailedEventAttributes_cause - The cause of the failure. This information is generated by the system and can be useful for diagnostic purposes.

If cause is set to OPERATION_NOT_PERMITTED, the decision failed because it lacked sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

$sel:decisionTaskCompletedEventId:CancelTimerFailedEventAttributes', cancelTimerFailedEventAttributes_decisionTaskCompletedEventId - The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the CancelTimer decision to cancel this timer. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

cancelTimerFailedEventAttributes_timerId :: Lens' CancelTimerFailedEventAttributes Text

The timerId provided in the CancelTimer decision that failed.

cancelTimerFailedEventAttributes_cause :: Lens' CancelTimerFailedEventAttributes CancelTimerFailedCause

The cause of the failure. This information is generated by the system and can be useful for diagnostic purposes.

If cause is set to OPERATION_NOT_PERMITTED, the decision failed because it lacked sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

cancelTimerFailedEventAttributes_decisionTaskCompletedEventId :: Lens' CancelTimerFailedEventAttributes Integer

The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the CancelTimer decision to cancel this timer. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.