Network.AWS.SWF.Types.ChildWorkflowExecutionCompletedEventAttributes

Description

Documentation

data ChildWorkflowExecutionCompletedEventAttributes

Provides the details of the ChildWorkflowExecutionCompleted event.

See: newChildWorkflowExecutionCompletedEventAttributes smart constructor.

Constructors

ChildWorkflowExecutionCompletedEventAttributes'

Fields

result :: Maybe Text

The result of the child workflow execution.

workflowExecution :: WorkflowExecution

The child workflow execution that was completed.

workflowType :: WorkflowType

The type of the child workflow execution.

initiatedEventId :: Integer

The ID of the StartChildWorkflowExecutionInitiated event corresponding to the StartChildWorkflowExecution Decision to start this child workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

startedEventId :: Integer

The ID of the ChildWorkflowExecutionStarted event recorded when this child workflow execution was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

newChildWorkflowExecutionCompletedEventAttributes

Arguments

:: WorkflowExecution
$sel:workflowExecution:ChildWorkflowExecutionCompletedEventAttributes'

-> WorkflowType
$sel:workflowType:ChildWorkflowExecutionCompletedEventAttributes'

-> Integer
$sel:initiatedEventId:ChildWorkflowExecutionCompletedEventAttributes'

-> Integer
$sel:startedEventId:ChildWorkflowExecutionCompletedEventAttributes'

-> ChildWorkflowExecutionCompletedEventAttributes

Create a value of ChildWorkflowExecutionCompletedEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:result:ChildWorkflowExecutionCompletedEventAttributes', childWorkflowExecutionCompletedEventAttributes_result - The result of the child workflow execution.

$sel:workflowExecution:ChildWorkflowExecutionCompletedEventAttributes', childWorkflowExecutionCompletedEventAttributes_workflowExecution - The child workflow execution that was completed.

$sel:workflowType:ChildWorkflowExecutionCompletedEventAttributes', childWorkflowExecutionCompletedEventAttributes_workflowType - The type of the child workflow execution.

$sel:initiatedEventId:ChildWorkflowExecutionCompletedEventAttributes', childWorkflowExecutionCompletedEventAttributes_initiatedEventId - The ID of the StartChildWorkflowExecutionInitiated event corresponding to the StartChildWorkflowExecution Decision to start this child workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

$sel:startedEventId:ChildWorkflowExecutionCompletedEventAttributes', childWorkflowExecutionCompletedEventAttributes_startedEventId - The ID of the ChildWorkflowExecutionStarted event recorded when this child workflow execution was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

childWorkflowExecutionCompletedEventAttributes_result :: Lens' ChildWorkflowExecutionCompletedEventAttributes (Maybe Text)

The result of the child workflow execution.

childWorkflowExecutionCompletedEventAttributes_workflowExecution :: Lens' ChildWorkflowExecutionCompletedEventAttributes WorkflowExecution

The child workflow execution that was completed.

childWorkflowExecutionCompletedEventAttributes_workflowType :: Lens' ChildWorkflowExecutionCompletedEventAttributes WorkflowType

The type of the child workflow execution.

childWorkflowExecutionCompletedEventAttributes_initiatedEventId :: Lens' ChildWorkflowExecutionCompletedEventAttributes Integer

The ID of the StartChildWorkflowExecutionInitiated event corresponding to the StartChildWorkflowExecution Decision to start this child workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

childWorkflowExecutionCompletedEventAttributes_startedEventId :: Lens' ChildWorkflowExecutionCompletedEventAttributes Integer

The ID of the ChildWorkflowExecutionStarted event recorded when this child workflow execution was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.