Network.AWS.SWF.Types.CloseStatusFilter

Description

Documentation

data CloseStatusFilter

Used to filter the closed workflow executions in visibility APIs by their close status.

See: newCloseStatusFilter smart constructor.

Constructors

CloseStatusFilter'

Fields

status :: CloseStatus

The close status that must match the close status of an execution for it to meet the criteria of this filter.

newCloseStatusFilter

Arguments

:: CloseStatus
$sel:status:CloseStatusFilter'

-> CloseStatusFilter

Create a value of CloseStatusFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CloseStatusFilter', closeStatusFilter_status - The close status that must match the close status of an execution for it to meet the criteria of this filter.

closeStatusFilter_status :: Lens' CloseStatusFilter CloseStatus

The close status that must match the close status of an execution for it to meet the criteria of this filter.