Network.AWS.SWF.Types.Decision

Description

Documentation

data Decision

Specifies a decision made by the decider. A decision can be one of these types:

CancelTimer – Cancels a previously started timer and records a TimerCanceled event in the history.

CancelWorkflowExecution – Closes the workflow execution and records a WorkflowExecutionCanceled event in the history.

CompleteWorkflowExecution – Closes the workflow execution and records a WorkflowExecutionCompleted event in the history .

ContinueAsNewWorkflowExecution – Closes the workflow execution and starts a new workflow execution of the same type using the same workflow ID and a unique run Id. A WorkflowExecutionContinuedAsNew event is recorded in the history.

FailWorkflowExecution – Closes the workflow execution and records a WorkflowExecutionFailed event in the history.

RecordMarker – Records a MarkerRecorded event in the history. Markers can be used for adding custom information in the history for instance to let deciders know that they don't need to look at the history beyond the marker event.

RequestCancelActivityTask – Attempts to cancel a previously scheduled activity task. If the activity task was scheduled but has not been assigned to a worker, then it is canceled. If the activity task was already assigned to a worker, then the worker is informed that cancellation has been requested in the response to RecordActivityTaskHeartbeat.

RequestCancelExternalWorkflowExecution – Requests that a request be made to cancel the specified external workflow execution and records a RequestCancelExternalWorkflowExecutionInitiated event in the history.

ScheduleActivityTask – Schedules an activity task.

SignalExternalWorkflowExecution – Requests a signal to be delivered to the specified external workflow execution and records a SignalExternalWorkflowExecutionInitiated event in the history.

StartChildWorkflowExecution – Requests that a child workflow execution be started and records a StartChildWorkflowExecutionInitiated event in the history. The child workflow execution is a separate workflow execution with its own history.

StartTimer – Starts a timer for this workflow execution and records a TimerStarted event in the history. This timer fires after the specified delay and record a TimerFired event.

Access Control

If you grant permission to use RespondDecisionTaskCompleted, you can use IAM policies to express permissions for the list of decisions returned by this action as if they were members of the API. Treating decisions as a pseudo API maintains a uniform conceptual model and helps keep policies readable. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

Decision Failure

Decisions can fail for several reasons

The ordering of decisions should follow a logical flow. Some decisions might not make sense in the current context of the workflow execution and therefore fails.

A limit on your account was reached.

The decision lacks sufficient permissions.

One of the following events might be added to the history to indicate an error. The event attribute's cause parameter indicates the cause. If cause is set to OPERATION_NOT_PERMITTED, the decision failed because it lacked sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

ScheduleActivityTaskFailed – A ScheduleActivityTask decision failed. This could happen if the activity type specified in the decision isn't registered, is in a deprecated state, or the decision isn't properly configured.

RequestCancelActivityTaskFailed – A RequestCancelActivityTask decision failed. This could happen if there is no open activity task with the specified activityId.

StartTimerFailed – A StartTimer decision failed. This could happen if there is another open timer with the same timerId.

CancelTimerFailed – A CancelTimer decision failed. This could happen if there is no open timer with the specified timerId.

StartChildWorkflowExecutionFailed – A StartChildWorkflowExecution decision failed. This could happen if the workflow type specified isn't registered, is deprecated, or the decision isn't properly configured.

SignalExternalWorkflowExecutionFailed – A SignalExternalWorkflowExecution decision failed. This could happen if the workflowID specified in the decision was incorrect.

RequestCancelExternalWorkflowExecutionFailed – A RequestCancelExternalWorkflowExecution decision failed. This could happen if the workflowID specified in the decision was incorrect.

CancelWorkflowExecutionFailed – A CancelWorkflowExecution decision failed. This could happen if there is an unhandled decision task pending in the workflow execution.

CompleteWorkflowExecutionFailed – A CompleteWorkflowExecution decision failed. This could happen if there is an unhandled decision task pending in the workflow execution.

ContinueAsNewWorkflowExecutionFailed – A ContinueAsNewWorkflowExecution decision failed. This could happen if there is an unhandled decision task pending in the workflow execution or the ContinueAsNewWorkflowExecution decision was not configured correctly.

FailWorkflowExecutionFailed – A FailWorkflowExecution decision failed. This could happen if there is an unhandled decision task pending in the workflow execution.

The preceding error events might occur due to an error in the decider logic, which might put the workflow execution in an unstable state The cause field in the event structure for the error event indicates the cause of the error.

A workflow execution may be closed by the decider by returning one of the following decisions when completing a decision task: CompleteWorkflowExecution, FailWorkflowExecution, CancelWorkflowExecution and ContinueAsNewWorkflowExecution. An UnhandledDecision fault is returned if a workflow closing decision is specified and a signal or activity event had been added to the history while the decision task was being performed by the decider. Unlike the above situations which are logic issues, this fault is always possible because of race conditions in a distributed system. The right action here is to call RespondDecisionTaskCompleted without any decisions. This would result in another decision task with these new events included in the history. The decider should handle the new events and may decide to close the workflow execution.

How to Code a Decision

You code a decision by first setting the decision type field to one of the above decision values, and then set the corresponding attributes field shown below:

 ScheduleActivityTaskDecisionAttributes

 RequestCancelActivityTaskDecisionAttributes

 CompleteWorkflowExecutionDecisionAttributes

 FailWorkflowExecutionDecisionAttributes

 CancelWorkflowExecutionDecisionAttributes

 ContinueAsNewWorkflowExecutionDecisionAttributes

 RecordMarkerDecisionAttributes

 StartTimerDecisionAttributes

 CancelTimerDecisionAttributes

 SignalExternalWorkflowExecutionDecisionAttributes

 RequestCancelExternalWorkflowExecutionDecisionAttributes

 StartChildWorkflowExecutionDecisionAttributes

See: newDecision smart constructor.

Constructors

Decision'

Fields

completeWorkflowExecutionDecisionAttributes :: Maybe CompleteWorkflowExecutionDecisionAttributes

Provides the details of the CompleteWorkflowExecution decision. It isn't set for other decision types.

failWorkflowExecutionDecisionAttributes :: Maybe FailWorkflowExecutionDecisionAttributes

Provides the details of the FailWorkflowExecution decision. It isn't set for other decision types.

startChildWorkflowExecutionDecisionAttributes :: Maybe StartChildWorkflowExecutionDecisionAttributes

Provides the details of the StartChildWorkflowExecution decision. It isn't set for other decision types.

requestCancelExternalWorkflowExecutionDecisionAttributes :: Maybe RequestCancelExternalWorkflowExecutionDecisionAttributes

Provides the details of the RequestCancelExternalWorkflowExecution decision. It isn't set for other decision types.

recordMarkerDecisionAttributes :: Maybe RecordMarkerDecisionAttributes

Provides the details of the RecordMarker decision. It isn't set for other decision types.

cancelWorkflowExecutionDecisionAttributes :: Maybe CancelWorkflowExecutionDecisionAttributes

Provides the details of the CancelWorkflowExecution decision. It isn't set for other decision types.

requestCancelActivityTaskDecisionAttributes :: Maybe RequestCancelActivityTaskDecisionAttributes

Provides the details of the RequestCancelActivityTask decision. It isn't set for other decision types.

startTimerDecisionAttributes :: Maybe StartTimerDecisionAttributes

Provides the details of the StartTimer decision. It isn't set for other decision types.

signalExternalWorkflowExecutionDecisionAttributes :: Maybe SignalExternalWorkflowExecutionDecisionAttributes

Provides the details of the SignalExternalWorkflowExecution decision. It isn't set for other decision types.

scheduleActivityTaskDecisionAttributes :: Maybe ScheduleActivityTaskDecisionAttributes

Provides the details of the ScheduleActivityTask decision. It isn't set for other decision types.

scheduleLambdaFunctionDecisionAttributes :: Maybe ScheduleLambdaFunctionDecisionAttributes

Provides the details of the ScheduleLambdaFunction decision. It isn't set for other decision types.

continueAsNewWorkflowExecutionDecisionAttributes :: Maybe ContinueAsNewWorkflowExecutionDecisionAttributes

Provides the details of the ContinueAsNewWorkflowExecution decision. It isn't set for other decision types.

cancelTimerDecisionAttributes :: Maybe CancelTimerDecisionAttributes

Provides the details of the CancelTimer decision. It isn't set for other decision types.

decisionType :: DecisionType

Specifies the type of the decision.

newDecision

Arguments

:: DecisionType
$sel:decisionType:Decision'

-> Decision

Create a value of Decision with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:completeWorkflowExecutionDecisionAttributes:Decision', decision_completeWorkflowExecutionDecisionAttributes - Provides the details of the CompleteWorkflowExecution decision. It isn't set for other decision types.

$sel:failWorkflowExecutionDecisionAttributes:Decision', decision_failWorkflowExecutionDecisionAttributes - Provides the details of the FailWorkflowExecution decision. It isn't set for other decision types.

$sel:startChildWorkflowExecutionDecisionAttributes:Decision', decision_startChildWorkflowExecutionDecisionAttributes - Provides the details of the StartChildWorkflowExecution decision. It isn't set for other decision types.

$sel:requestCancelExternalWorkflowExecutionDecisionAttributes:Decision', decision_requestCancelExternalWorkflowExecutionDecisionAttributes - Provides the details of the RequestCancelExternalWorkflowExecution decision. It isn't set for other decision types.

$sel:recordMarkerDecisionAttributes:Decision', decision_recordMarkerDecisionAttributes - Provides the details of the RecordMarker decision. It isn't set for other decision types.

$sel:cancelWorkflowExecutionDecisionAttributes:Decision', decision_cancelWorkflowExecutionDecisionAttributes - Provides the details of the CancelWorkflowExecution decision. It isn't set for other decision types.

$sel:requestCancelActivityTaskDecisionAttributes:Decision', decision_requestCancelActivityTaskDecisionAttributes - Provides the details of the RequestCancelActivityTask decision. It isn't set for other decision types.

$sel:startTimerDecisionAttributes:Decision', decision_startTimerDecisionAttributes - Provides the details of the StartTimer decision. It isn't set for other decision types.

$sel:signalExternalWorkflowExecutionDecisionAttributes:Decision', decision_signalExternalWorkflowExecutionDecisionAttributes - Provides the details of the SignalExternalWorkflowExecution decision. It isn't set for other decision types.

$sel:scheduleActivityTaskDecisionAttributes:Decision', decision_scheduleActivityTaskDecisionAttributes - Provides the details of the ScheduleActivityTask decision. It isn't set for other decision types.

$sel:scheduleLambdaFunctionDecisionAttributes:Decision', decision_scheduleLambdaFunctionDecisionAttributes - Provides the details of the ScheduleLambdaFunction decision. It isn't set for other decision types.

$sel:continueAsNewWorkflowExecutionDecisionAttributes:Decision', decision_continueAsNewWorkflowExecutionDecisionAttributes - Provides the details of the ContinueAsNewWorkflowExecution decision. It isn't set for other decision types.

$sel:cancelTimerDecisionAttributes:Decision', decision_cancelTimerDecisionAttributes - Provides the details of the CancelTimer decision. It isn't set for other decision types.

$sel:decisionType:Decision', decision_decisionType - Specifies the type of the decision.

decision_completeWorkflowExecutionDecisionAttributes :: Lens' Decision (Maybe CompleteWorkflowExecutionDecisionAttributes)

Provides the details of the CompleteWorkflowExecution decision. It isn't set for other decision types.

decision_failWorkflowExecutionDecisionAttributes :: Lens' Decision (Maybe FailWorkflowExecutionDecisionAttributes)

Provides the details of the FailWorkflowExecution decision. It isn't set for other decision types.

decision_startChildWorkflowExecutionDecisionAttributes :: Lens' Decision (Maybe StartChildWorkflowExecutionDecisionAttributes)

Provides the details of the StartChildWorkflowExecution decision. It isn't set for other decision types.

decision_requestCancelExternalWorkflowExecutionDecisionAttributes :: Lens' Decision (Maybe RequestCancelExternalWorkflowExecutionDecisionAttributes)

Provides the details of the RequestCancelExternalWorkflowExecution decision. It isn't set for other decision types.

decision_recordMarkerDecisionAttributes :: Lens' Decision (Maybe RecordMarkerDecisionAttributes)

Provides the details of the RecordMarker decision. It isn't set for other decision types.

decision_cancelWorkflowExecutionDecisionAttributes :: Lens' Decision (Maybe CancelWorkflowExecutionDecisionAttributes)

Provides the details of the CancelWorkflowExecution decision. It isn't set for other decision types.

decision_requestCancelActivityTaskDecisionAttributes :: Lens' Decision (Maybe RequestCancelActivityTaskDecisionAttributes)

Provides the details of the RequestCancelActivityTask decision. It isn't set for other decision types.

decision_startTimerDecisionAttributes :: Lens' Decision (Maybe StartTimerDecisionAttributes)

Provides the details of the StartTimer decision. It isn't set for other decision types.

decision_signalExternalWorkflowExecutionDecisionAttributes :: Lens' Decision (Maybe SignalExternalWorkflowExecutionDecisionAttributes)

Provides the details of the SignalExternalWorkflowExecution decision. It isn't set for other decision types.

decision_scheduleActivityTaskDecisionAttributes :: Lens' Decision (Maybe ScheduleActivityTaskDecisionAttributes)

Provides the details of the ScheduleActivityTask decision. It isn't set for other decision types.

decision_scheduleLambdaFunctionDecisionAttributes :: Lens' Decision (Maybe ScheduleLambdaFunctionDecisionAttributes)

Provides the details of the ScheduleLambdaFunction decision. It isn't set for other decision types.

decision_continueAsNewWorkflowExecutionDecisionAttributes :: Lens' Decision (Maybe ContinueAsNewWorkflowExecutionDecisionAttributes)

Provides the details of the ContinueAsNewWorkflowExecution decision. It isn't set for other decision types.

decision_cancelTimerDecisionAttributes :: Lens' Decision (Maybe CancelTimerDecisionAttributes)

Provides the details of the CancelTimer decision. It isn't set for other decision types.

decision_decisionType :: Lens' Decision DecisionType

Specifies the type of the decision.