Network.AWS.SWF.Types.DomainInfo

Description

Documentation

data DomainInfo

Contains general information about a domain.

See: newDomainInfo smart constructor.

Constructors

DomainInfo'

Fields

arn :: Maybe Text

The ARN of the domain.

description :: Maybe Text

The description of the domain provided through RegisterDomain.

name :: Text

The name of the domain. This name is unique within the account.

status :: RegistrationStatus

The status of the domain:

REGISTERED – The domain is properly registered and available. You can use this domain for registering types and creating new workflow executions.

DEPRECATED – The domain was deprecated using DeprecateDomain, but is still in use. You should not create new workflow executions in this domain.

newDomainInfo

Arguments

:: Text
$sel:name:DomainInfo'

-> RegistrationStatus
$sel:status:DomainInfo'

-> DomainInfo

Create a value of DomainInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:DomainInfo', domainInfo_arn - The ARN of the domain.

$sel:description:DomainInfo', domainInfo_description - The description of the domain provided through RegisterDomain.

$sel:name:DomainInfo', domainInfo_name - The name of the domain. This name is unique within the account.

$sel:status:DomainInfo', domainInfo_status - The status of the domain:

REGISTERED – The domain is properly registered and available. You can use this domain for registering types and creating new workflow executions.

DEPRECATED – The domain was deprecated using DeprecateDomain, but is still in use. You should not create new workflow executions in this domain.

domainInfo_arn :: Lens' DomainInfo (Maybe Text)

The ARN of the domain.

domainInfo_description :: Lens' DomainInfo (Maybe Text)

The description of the domain provided through RegisterDomain.

domainInfo_name :: Lens' DomainInfo Text

The name of the domain. This name is unique within the account.

domainInfo_status :: Lens' DomainInfo RegistrationStatus

The status of the domain:

REGISTERED – The domain is properly registered and available. You can use this domain for registering types and creating new workflow executions.

DEPRECATED – The domain was deprecated using DeprecateDomain, but is still in use. You should not create new workflow executions in this domain.