Network.AWS.SWF.Types.ExternalWorkflowExecutionCancelRequestedEventAttributes

Description

Documentation

data ExternalWorkflowExecutionCancelRequestedEventAttributes

Provides the details of the ExternalWorkflowExecutionCancelRequested event.

See: newExternalWorkflowExecutionCancelRequestedEventAttributes smart constructor.

Constructors

ExternalWorkflowExecutionCancelRequestedEventAttributes'

Fields

workflowExecution :: WorkflowExecution

The external workflow execution to which the cancellation request was delivered.

initiatedEventId :: Integer

The ID of the RequestCancelExternalWorkflowExecutionInitiated event corresponding to the RequestCancelExternalWorkflowExecution decision to cancel this external workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

newExternalWorkflowExecutionCancelRequestedEventAttributes

Arguments

:: WorkflowExecution
$sel:workflowExecution:ExternalWorkflowExecutionCancelRequestedEventAttributes'

-> Integer
$sel:initiatedEventId:ExternalWorkflowExecutionCancelRequestedEventAttributes'

-> ExternalWorkflowExecutionCancelRequestedEventAttributes

Create a value of ExternalWorkflowExecutionCancelRequestedEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workflowExecution:ExternalWorkflowExecutionCancelRequestedEventAttributes', externalWorkflowExecutionCancelRequestedEventAttributes_workflowExecution - The external workflow execution to which the cancellation request was delivered.

$sel:initiatedEventId:ExternalWorkflowExecutionCancelRequestedEventAttributes', externalWorkflowExecutionCancelRequestedEventAttributes_initiatedEventId - The ID of the RequestCancelExternalWorkflowExecutionInitiated event corresponding to the RequestCancelExternalWorkflowExecution decision to cancel this external workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

externalWorkflowExecutionCancelRequestedEventAttributes_workflowExecution :: Lens' ExternalWorkflowExecutionCancelRequestedEventAttributes WorkflowExecution

The external workflow execution to which the cancellation request was delivered.

externalWorkflowExecutionCancelRequestedEventAttributes_initiatedEventId :: Lens' ExternalWorkflowExecutionCancelRequestedEventAttributes Integer

The ID of the RequestCancelExternalWorkflowExecutionInitiated event corresponding to the RequestCancelExternalWorkflowExecution decision to cancel this external workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.