Network.AWS.SWF.Types.HistoryEvent

Description

Documentation

data HistoryEvent

Event within a workflow execution. A history event can be one of these types:

ActivityTaskCancelRequested – A RequestCancelActivityTask decision was received by the system.

ActivityTaskCanceled – The activity task was successfully canceled.

ActivityTaskCompleted – An activity worker successfully completed an activity task by calling RespondActivityTaskCompleted.

ActivityTaskFailed – An activity worker failed an activity task by calling RespondActivityTaskFailed.

ActivityTaskScheduled – An activity task was scheduled for execution.

ActivityTaskStarted – The scheduled activity task was dispatched to a worker.

ActivityTaskTimedOut – The activity task timed out.

CancelTimerFailed – Failed to process CancelTimer decision. This happens when the decision isn't configured properly, for example no timer exists with the specified timer Id.

CancelWorkflowExecutionFailed – A request to cancel a workflow execution failed.

ChildWorkflowExecutionCanceled – A child workflow execution, started by this workflow execution, was canceled and closed.

ChildWorkflowExecutionCompleted – A child workflow execution, started by this workflow execution, completed successfully and was closed.

ChildWorkflowExecutionFailed – A child workflow execution, started by this workflow execution, failed to complete successfully and was closed.

ChildWorkflowExecutionStarted – A child workflow execution was successfully started.

ChildWorkflowExecutionTerminated – A child workflow execution, started by this workflow execution, was terminated.

ChildWorkflowExecutionTimedOut – A child workflow execution, started by this workflow execution, timed out and was closed.

CompleteWorkflowExecutionFailed – The workflow execution failed to complete.

ContinueAsNewWorkflowExecutionFailed – The workflow execution failed to complete after being continued as a new workflow execution.

DecisionTaskCompleted – The decider successfully completed a decision task by calling RespondDecisionTaskCompleted.

DecisionTaskScheduled – A decision task was scheduled for the workflow execution.

DecisionTaskStarted – The decision task was dispatched to a decider.

DecisionTaskTimedOut – The decision task timed out.

ExternalWorkflowExecutionCancelRequested – Request to cancel an external workflow execution was successfully delivered to the target execution.

ExternalWorkflowExecutionSignaled – A signal, requested by this workflow execution, was successfully delivered to the target external workflow execution.

FailWorkflowExecutionFailed – A request to mark a workflow execution as failed, itself failed.

MarkerRecorded – A marker was recorded in the workflow history as the result of a RecordMarker decision.

RecordMarkerFailed – A RecordMarker decision was returned as failed.

RequestCancelActivityTaskFailed – Failed to process RequestCancelActivityTask decision. This happens when the decision isn't configured properly.

RequestCancelExternalWorkflowExecutionFailed – Request to cancel an external workflow execution failed.

RequestCancelExternalWorkflowExecutionInitiated – A request was made to request the cancellation of an external workflow execution.

ScheduleActivityTaskFailed – Failed to process ScheduleActivityTask decision. This happens when the decision isn't configured properly, for example the activity type specified isn't registered.

SignalExternalWorkflowExecutionFailed – The request to signal an external workflow execution failed.

SignalExternalWorkflowExecutionInitiated – A request to signal an external workflow was made.

StartActivityTaskFailed – A scheduled activity task failed to start.

StartChildWorkflowExecutionFailed – Failed to process StartChildWorkflowExecution decision. This happens when the decision isn't configured properly, for example the workflow type specified isn't registered.

StartChildWorkflowExecutionInitiated – A request was made to start a child workflow execution.

StartTimerFailed – Failed to process StartTimer decision. This happens when the decision isn't configured properly, for example a timer already exists with the specified timer Id.

TimerCanceled – A timer, previously started for this workflow execution, was successfully canceled.

TimerFired – A timer, previously started for this workflow execution, fired.

TimerStarted – A timer was started for the workflow execution due to a StartTimer decision.

WorkflowExecutionCancelRequested – A request to cancel this workflow execution was made.

WorkflowExecutionCanceled – The workflow execution was successfully canceled and closed.

WorkflowExecutionCompleted – The workflow execution was closed due to successful completion.

WorkflowExecutionContinuedAsNew – The workflow execution was closed and a new execution of the same type was created with the same workflowId.

WorkflowExecutionFailed – The workflow execution closed due to a failure.

WorkflowExecutionSignaled – An external signal was received for the workflow execution.

WorkflowExecutionStarted – The workflow execution was started.

WorkflowExecutionTerminated – The workflow execution was terminated.

WorkflowExecutionTimedOut – The workflow execution was closed because a time out was exceeded.

See: newHistoryEvent smart constructor.

Constructors

HistoryEvent'

Fields

workflowExecutionCancelRequestedEventAttributes :: Maybe WorkflowExecutionCancelRequestedEventAttributes

If the event is of type WorkflowExecutionCancelRequested then this member is set and provides detailed information about the event. It isn't set for other event types.

recordMarkerFailedEventAttributes :: Maybe RecordMarkerFailedEventAttributes

If the event is of type DecisionTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

requestCancelExternalWorkflowExecutionInitiatedEventAttributes :: Maybe RequestCancelExternalWorkflowExecutionInitiatedEventAttributes

If the event is of type RequestCancelExternalWorkflowExecutionInitiated then this member is set and provides detailed information about the event. It isn't set for other event types.

lambdaFunctionStartedEventAttributes :: Maybe LambdaFunctionStartedEventAttributes

Provides the details of the LambdaFunctionStarted event. It isn't set for other event types.

decisionTaskScheduledEventAttributes :: Maybe DecisionTaskScheduledEventAttributes

If the event is of type DecisionTaskScheduled then this member is set and provides detailed information about the event. It isn't set for other event types.

workflowExecutionCompletedEventAttributes :: Maybe WorkflowExecutionCompletedEventAttributes

If the event is of type WorkflowExecutionCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

startTimerFailedEventAttributes :: Maybe StartTimerFailedEventAttributes

If the event is of type StartTimerFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

activityTaskScheduledEventAttributes :: Maybe ActivityTaskScheduledEventAttributes

If the event is of type ActivityTaskScheduled then this member is set and provides detailed information about the event. It isn't set for other event types.

scheduleActivityTaskFailedEventAttributes :: Maybe ScheduleActivityTaskFailedEventAttributes

If the event is of type ScheduleActivityTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

childWorkflowExecutionCompletedEventAttributes :: Maybe ChildWorkflowExecutionCompletedEventAttributes

If the event is of type ChildWorkflowExecutionCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

markerRecordedEventAttributes :: Maybe MarkerRecordedEventAttributes

If the event is of type MarkerRecorded then this member is set and provides detailed information about the event. It isn't set for other event types.

scheduleLambdaFunctionFailedEventAttributes :: Maybe ScheduleLambdaFunctionFailedEventAttributes

Provides the details of the ScheduleLambdaFunctionFailed event. It isn't set for other event types.

completeWorkflowExecutionFailedEventAttributes :: Maybe CompleteWorkflowExecutionFailedEventAttributes

If the event is of type CompleteWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

lambdaFunctionCompletedEventAttributes :: Maybe LambdaFunctionCompletedEventAttributes

Provides the details of the LambdaFunctionCompleted event. It isn't set for other event types.

requestCancelExternalWorkflowExecutionFailedEventAttributes :: Maybe RequestCancelExternalWorkflowExecutionFailedEventAttributes

If the event is of type RequestCancelExternalWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

timerCanceledEventAttributes :: Maybe TimerCanceledEventAttributes

If the event is of type TimerCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

workflowExecutionStartedEventAttributes :: Maybe WorkflowExecutionStartedEventAttributes

If the event is of type WorkflowExecutionStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

activityTaskCompletedEventAttributes :: Maybe ActivityTaskCompletedEventAttributes

If the event is of type ActivityTaskCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

decisionTaskTimedOutEventAttributes :: Maybe DecisionTaskTimedOutEventAttributes

If the event is of type DecisionTaskTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

cancelTimerFailedEventAttributes :: Maybe CancelTimerFailedEventAttributes

If the event is of type CancelTimerFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

childWorkflowExecutionStartedEventAttributes :: Maybe ChildWorkflowExecutionStartedEventAttributes

If the event is of type ChildWorkflowExecutionStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

activityTaskCanceledEventAttributes :: Maybe ActivityTaskCanceledEventAttributes

If the event is of type ActivityTaskCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

activityTaskTimedOutEventAttributes :: Maybe ActivityTaskTimedOutEventAttributes

If the event is of type ActivityTaskTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

decisionTaskStartedEventAttributes :: Maybe DecisionTaskStartedEventAttributes

If the event is of type DecisionTaskStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

workflowExecutionTerminatedEventAttributes :: Maybe WorkflowExecutionTerminatedEventAttributes

If the event is of type WorkflowExecutionTerminated then this member is set and provides detailed information about the event. It isn't set for other event types.

childWorkflowExecutionCanceledEventAttributes :: Maybe ChildWorkflowExecutionCanceledEventAttributes

If the event is of type ChildWorkflowExecutionCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

requestCancelActivityTaskFailedEventAttributes :: Maybe RequestCancelActivityTaskFailedEventAttributes

If the event is of type RequestCancelActivityTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

lambdaFunctionScheduledEventAttributes :: Maybe LambdaFunctionScheduledEventAttributes

Provides the details of the LambdaFunctionScheduled event. It isn't set for other event types.

childWorkflowExecutionTimedOutEventAttributes :: Maybe ChildWorkflowExecutionTimedOutEventAttributes

If the event is of type ChildWorkflowExecutionTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

cancelWorkflowExecutionFailedEventAttributes :: Maybe CancelWorkflowExecutionFailedEventAttributes

If the event is of type CancelWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

startChildWorkflowExecutionInitiatedEventAttributes :: Maybe StartChildWorkflowExecutionInitiatedEventAttributes

If the event is of type StartChildWorkflowExecutionInitiated then this member is set and provides detailed information about the event. It isn't set for other event types.

signalExternalWorkflowExecutionFailedEventAttributes :: Maybe SignalExternalWorkflowExecutionFailedEventAttributes

If the event is of type SignalExternalWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

activityTaskStartedEventAttributes :: Maybe ActivityTaskStartedEventAttributes

If the event is of type ActivityTaskStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

startLambdaFunctionFailedEventAttributes :: Maybe StartLambdaFunctionFailedEventAttributes

Provides the details of the StartLambdaFunctionFailed event. It isn't set for other event types.

childWorkflowExecutionTerminatedEventAttributes :: Maybe ChildWorkflowExecutionTerminatedEventAttributes

If the event is of type ChildWorkflowExecutionTerminated then this member is set and provides detailed information about the event. It isn't set for other event types.

lambdaFunctionFailedEventAttributes :: Maybe LambdaFunctionFailedEventAttributes

Provides the details of the LambdaFunctionFailed event. It isn't set for other event types.

workflowExecutionCanceledEventAttributes :: Maybe WorkflowExecutionCanceledEventAttributes

If the event is of type WorkflowExecutionCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

timerStartedEventAttributes :: Maybe TimerStartedEventAttributes

If the event is of type TimerStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

activityTaskCancelRequestedEventAttributes :: Maybe ActivityTaskCancelRequestedEventAttributes

If the event is of type ActivityTaskcancelRequested then this member is set and provides detailed information about the event. It isn't set for other event types.

workflowExecutionTimedOutEventAttributes :: Maybe WorkflowExecutionTimedOutEventAttributes

If the event is of type WorkflowExecutionTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

workflowExecutionSignaledEventAttributes :: Maybe WorkflowExecutionSignaledEventAttributes

If the event is of type WorkflowExecutionSignaled then this member is set and provides detailed information about the event. It isn't set for other event types.

timerFiredEventAttributes :: Maybe TimerFiredEventAttributes

If the event is of type TimerFired then this member is set and provides detailed information about the event. It isn't set for other event types.

activityTaskFailedEventAttributes :: Maybe ActivityTaskFailedEventAttributes

If the event is of type ActivityTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

externalWorkflowExecutionSignaledEventAttributes :: Maybe ExternalWorkflowExecutionSignaledEventAttributes

If the event is of type ExternalWorkflowExecutionSignaled then this member is set and provides detailed information about the event. It isn't set for other event types.

decisionTaskCompletedEventAttributes :: Maybe DecisionTaskCompletedEventAttributes

If the event is of type DecisionTaskCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

startChildWorkflowExecutionFailedEventAttributes :: Maybe StartChildWorkflowExecutionFailedEventAttributes

If the event is of type StartChildWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

childWorkflowExecutionFailedEventAttributes :: Maybe ChildWorkflowExecutionFailedEventAttributes

If the event is of type ChildWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

failWorkflowExecutionFailedEventAttributes :: Maybe FailWorkflowExecutionFailedEventAttributes

If the event is of type FailWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

continueAsNewWorkflowExecutionFailedEventAttributes :: Maybe ContinueAsNewWorkflowExecutionFailedEventAttributes

If the event is of type ContinueAsNewWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

signalExternalWorkflowExecutionInitiatedEventAttributes :: Maybe SignalExternalWorkflowExecutionInitiatedEventAttributes

If the event is of type SignalExternalWorkflowExecutionInitiated then this member is set and provides detailed information about the event. It isn't set for other event types.

lambdaFunctionTimedOutEventAttributes :: Maybe LambdaFunctionTimedOutEventAttributes

Provides the details of the LambdaFunctionTimedOut event. It isn't set for other event types.

workflowExecutionFailedEventAttributes :: Maybe WorkflowExecutionFailedEventAttributes

If the event is of type WorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

workflowExecutionContinuedAsNewEventAttributes :: Maybe WorkflowExecutionContinuedAsNewEventAttributes

If the event is of type WorkflowExecutionContinuedAsNew then this member is set and provides detailed information about the event. It isn't set for other event types.

externalWorkflowExecutionCancelRequestedEventAttributes :: Maybe ExternalWorkflowExecutionCancelRequestedEventAttributes

If the event is of type ExternalWorkflowExecutionCancelRequested then this member is set and provides detailed information about the event. It isn't set for other event types.

eventTimestamp :: POSIX

The date and time when the event occurred.

eventType :: EventType

The type of the history event.

eventId :: Integer

The system generated ID of the event. This ID uniquely identifies the event with in the workflow execution history.

newHistoryEvent

Arguments

:: UTCTime
$sel:eventTimestamp:HistoryEvent'

-> EventType
$sel:eventType:HistoryEvent'

-> Integer
$sel:eventId:HistoryEvent'

-> HistoryEvent

Create a value of HistoryEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workflowExecutionCancelRequestedEventAttributes:HistoryEvent', historyEvent_workflowExecutionCancelRequestedEventAttributes - If the event is of type WorkflowExecutionCancelRequested then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:recordMarkerFailedEventAttributes:HistoryEvent', historyEvent_recordMarkerFailedEventAttributes - If the event is of type DecisionTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:requestCancelExternalWorkflowExecutionInitiatedEventAttributes:HistoryEvent', historyEvent_requestCancelExternalWorkflowExecutionInitiatedEventAttributes - If the event is of type RequestCancelExternalWorkflowExecutionInitiated then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:lambdaFunctionStartedEventAttributes:HistoryEvent', historyEvent_lambdaFunctionStartedEventAttributes - Provides the details of the LambdaFunctionStarted event. It isn't set for other event types.

$sel:decisionTaskScheduledEventAttributes:HistoryEvent', historyEvent_decisionTaskScheduledEventAttributes - If the event is of type DecisionTaskScheduled then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:workflowExecutionCompletedEventAttributes:HistoryEvent', historyEvent_workflowExecutionCompletedEventAttributes - If the event is of type WorkflowExecutionCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:startTimerFailedEventAttributes:HistoryEvent', historyEvent_startTimerFailedEventAttributes - If the event is of type StartTimerFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:activityTaskScheduledEventAttributes:HistoryEvent', historyEvent_activityTaskScheduledEventAttributes - If the event is of type ActivityTaskScheduled then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:scheduleActivityTaskFailedEventAttributes:HistoryEvent', historyEvent_scheduleActivityTaskFailedEventAttributes - If the event is of type ScheduleActivityTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:childWorkflowExecutionCompletedEventAttributes:HistoryEvent', historyEvent_childWorkflowExecutionCompletedEventAttributes - If the event is of type ChildWorkflowExecutionCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:markerRecordedEventAttributes:HistoryEvent', historyEvent_markerRecordedEventAttributes - If the event is of type MarkerRecorded then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:scheduleLambdaFunctionFailedEventAttributes:HistoryEvent', historyEvent_scheduleLambdaFunctionFailedEventAttributes - Provides the details of the ScheduleLambdaFunctionFailed event. It isn't set for other event types.

$sel:completeWorkflowExecutionFailedEventAttributes:HistoryEvent', historyEvent_completeWorkflowExecutionFailedEventAttributes - If the event is of type CompleteWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:lambdaFunctionCompletedEventAttributes:HistoryEvent', historyEvent_lambdaFunctionCompletedEventAttributes - Provides the details of the LambdaFunctionCompleted event. It isn't set for other event types.

$sel:requestCancelExternalWorkflowExecutionFailedEventAttributes:HistoryEvent', historyEvent_requestCancelExternalWorkflowExecutionFailedEventAttributes - If the event is of type RequestCancelExternalWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:timerCanceledEventAttributes:HistoryEvent', historyEvent_timerCanceledEventAttributes - If the event is of type TimerCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:workflowExecutionStartedEventAttributes:HistoryEvent', historyEvent_workflowExecutionStartedEventAttributes - If the event is of type WorkflowExecutionStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:activityTaskCompletedEventAttributes:HistoryEvent', historyEvent_activityTaskCompletedEventAttributes - If the event is of type ActivityTaskCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:decisionTaskTimedOutEventAttributes:HistoryEvent', historyEvent_decisionTaskTimedOutEventAttributes - If the event is of type DecisionTaskTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:cancelTimerFailedEventAttributes:HistoryEvent', historyEvent_cancelTimerFailedEventAttributes - If the event is of type CancelTimerFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:childWorkflowExecutionStartedEventAttributes:HistoryEvent', historyEvent_childWorkflowExecutionStartedEventAttributes - If the event is of type ChildWorkflowExecutionStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:activityTaskCanceledEventAttributes:HistoryEvent', historyEvent_activityTaskCanceledEventAttributes - If the event is of type ActivityTaskCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:activityTaskTimedOutEventAttributes:HistoryEvent', historyEvent_activityTaskTimedOutEventAttributes - If the event is of type ActivityTaskTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:decisionTaskStartedEventAttributes:HistoryEvent', historyEvent_decisionTaskStartedEventAttributes - If the event is of type DecisionTaskStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:workflowExecutionTerminatedEventAttributes:HistoryEvent', historyEvent_workflowExecutionTerminatedEventAttributes - If the event is of type WorkflowExecutionTerminated then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:childWorkflowExecutionCanceledEventAttributes:HistoryEvent', historyEvent_childWorkflowExecutionCanceledEventAttributes - If the event is of type ChildWorkflowExecutionCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:requestCancelActivityTaskFailedEventAttributes:HistoryEvent', historyEvent_requestCancelActivityTaskFailedEventAttributes - If the event is of type RequestCancelActivityTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:lambdaFunctionScheduledEventAttributes:HistoryEvent', historyEvent_lambdaFunctionScheduledEventAttributes - Provides the details of the LambdaFunctionScheduled event. It isn't set for other event types.

$sel:childWorkflowExecutionTimedOutEventAttributes:HistoryEvent', historyEvent_childWorkflowExecutionTimedOutEventAttributes - If the event is of type ChildWorkflowExecutionTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:cancelWorkflowExecutionFailedEventAttributes:HistoryEvent', historyEvent_cancelWorkflowExecutionFailedEventAttributes - If the event is of type CancelWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:startChildWorkflowExecutionInitiatedEventAttributes:HistoryEvent', historyEvent_startChildWorkflowExecutionInitiatedEventAttributes - If the event is of type StartChildWorkflowExecutionInitiated then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:signalExternalWorkflowExecutionFailedEventAttributes:HistoryEvent', historyEvent_signalExternalWorkflowExecutionFailedEventAttributes - If the event is of type SignalExternalWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:activityTaskStartedEventAttributes:HistoryEvent', historyEvent_activityTaskStartedEventAttributes - If the event is of type ActivityTaskStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:startLambdaFunctionFailedEventAttributes:HistoryEvent', historyEvent_startLambdaFunctionFailedEventAttributes - Provides the details of the StartLambdaFunctionFailed event. It isn't set for other event types.

$sel:childWorkflowExecutionTerminatedEventAttributes:HistoryEvent', historyEvent_childWorkflowExecutionTerminatedEventAttributes - If the event is of type ChildWorkflowExecutionTerminated then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:lambdaFunctionFailedEventAttributes:HistoryEvent', historyEvent_lambdaFunctionFailedEventAttributes - Provides the details of the LambdaFunctionFailed event. It isn't set for other event types.

$sel:workflowExecutionCanceledEventAttributes:HistoryEvent', historyEvent_workflowExecutionCanceledEventAttributes - If the event is of type WorkflowExecutionCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:timerStartedEventAttributes:HistoryEvent', historyEvent_timerStartedEventAttributes - If the event is of type TimerStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:activityTaskCancelRequestedEventAttributes:HistoryEvent', historyEvent_activityTaskCancelRequestedEventAttributes - If the event is of type ActivityTaskcancelRequested then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:workflowExecutionTimedOutEventAttributes:HistoryEvent', historyEvent_workflowExecutionTimedOutEventAttributes - If the event is of type WorkflowExecutionTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:workflowExecutionSignaledEventAttributes:HistoryEvent', historyEvent_workflowExecutionSignaledEventAttributes - If the event is of type WorkflowExecutionSignaled then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:timerFiredEventAttributes:HistoryEvent', historyEvent_timerFiredEventAttributes - If the event is of type TimerFired then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:activityTaskFailedEventAttributes:HistoryEvent', historyEvent_activityTaskFailedEventAttributes - If the event is of type ActivityTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:externalWorkflowExecutionSignaledEventAttributes:HistoryEvent', historyEvent_externalWorkflowExecutionSignaledEventAttributes - If the event is of type ExternalWorkflowExecutionSignaled then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:decisionTaskCompletedEventAttributes:HistoryEvent', historyEvent_decisionTaskCompletedEventAttributes - If the event is of type DecisionTaskCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:startChildWorkflowExecutionFailedEventAttributes:HistoryEvent', historyEvent_startChildWorkflowExecutionFailedEventAttributes - If the event is of type StartChildWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:childWorkflowExecutionFailedEventAttributes:HistoryEvent', historyEvent_childWorkflowExecutionFailedEventAttributes - If the event is of type ChildWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:failWorkflowExecutionFailedEventAttributes:HistoryEvent', historyEvent_failWorkflowExecutionFailedEventAttributes - If the event is of type FailWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:continueAsNewWorkflowExecutionFailedEventAttributes:HistoryEvent', historyEvent_continueAsNewWorkflowExecutionFailedEventAttributes - If the event is of type ContinueAsNewWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:signalExternalWorkflowExecutionInitiatedEventAttributes:HistoryEvent', historyEvent_signalExternalWorkflowExecutionInitiatedEventAttributes - If the event is of type SignalExternalWorkflowExecutionInitiated then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:lambdaFunctionTimedOutEventAttributes:HistoryEvent', historyEvent_lambdaFunctionTimedOutEventAttributes - Provides the details of the LambdaFunctionTimedOut event. It isn't set for other event types.

$sel:workflowExecutionFailedEventAttributes:HistoryEvent', historyEvent_workflowExecutionFailedEventAttributes - If the event is of type WorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:workflowExecutionContinuedAsNewEventAttributes:HistoryEvent', historyEvent_workflowExecutionContinuedAsNewEventAttributes - If the event is of type WorkflowExecutionContinuedAsNew then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:externalWorkflowExecutionCancelRequestedEventAttributes:HistoryEvent', historyEvent_externalWorkflowExecutionCancelRequestedEventAttributes - If the event is of type ExternalWorkflowExecutionCancelRequested then this member is set and provides detailed information about the event. It isn't set for other event types.

$sel:eventTimestamp:HistoryEvent', historyEvent_eventTimestamp - The date and time when the event occurred.

$sel:eventType:HistoryEvent', historyEvent_eventType - The type of the history event.

$sel:eventId:HistoryEvent', historyEvent_eventId - The system generated ID of the event. This ID uniquely identifies the event with in the workflow execution history.

historyEvent_workflowExecutionCancelRequestedEventAttributes :: Lens' HistoryEvent (Maybe WorkflowExecutionCancelRequestedEventAttributes)

If the event is of type WorkflowExecutionCancelRequested then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_recordMarkerFailedEventAttributes :: Lens' HistoryEvent (Maybe RecordMarkerFailedEventAttributes)

If the event is of type DecisionTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_requestCancelExternalWorkflowExecutionInitiatedEventAttributes :: Lens' HistoryEvent (Maybe RequestCancelExternalWorkflowExecutionInitiatedEventAttributes)

If the event is of type RequestCancelExternalWorkflowExecutionInitiated then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_lambdaFunctionStartedEventAttributes :: Lens' HistoryEvent (Maybe LambdaFunctionStartedEventAttributes)

Provides the details of the LambdaFunctionStarted event. It isn't set for other event types.

historyEvent_decisionTaskScheduledEventAttributes :: Lens' HistoryEvent (Maybe DecisionTaskScheduledEventAttributes)

If the event is of type DecisionTaskScheduled then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_workflowExecutionCompletedEventAttributes :: Lens' HistoryEvent (Maybe WorkflowExecutionCompletedEventAttributes)

If the event is of type WorkflowExecutionCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_startTimerFailedEventAttributes :: Lens' HistoryEvent (Maybe StartTimerFailedEventAttributes)

If the event is of type StartTimerFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_activityTaskScheduledEventAttributes :: Lens' HistoryEvent (Maybe ActivityTaskScheduledEventAttributes)

If the event is of type ActivityTaskScheduled then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_scheduleActivityTaskFailedEventAttributes :: Lens' HistoryEvent (Maybe ScheduleActivityTaskFailedEventAttributes)

If the event is of type ScheduleActivityTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_childWorkflowExecutionCompletedEventAttributes :: Lens' HistoryEvent (Maybe ChildWorkflowExecutionCompletedEventAttributes)

If the event is of type ChildWorkflowExecutionCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_markerRecordedEventAttributes :: Lens' HistoryEvent (Maybe MarkerRecordedEventAttributes)

If the event is of type MarkerRecorded then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_scheduleLambdaFunctionFailedEventAttributes :: Lens' HistoryEvent (Maybe ScheduleLambdaFunctionFailedEventAttributes)

Provides the details of the ScheduleLambdaFunctionFailed event. It isn't set for other event types.

historyEvent_completeWorkflowExecutionFailedEventAttributes :: Lens' HistoryEvent (Maybe CompleteWorkflowExecutionFailedEventAttributes)

If the event is of type CompleteWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_lambdaFunctionCompletedEventAttributes :: Lens' HistoryEvent (Maybe LambdaFunctionCompletedEventAttributes)

Provides the details of the LambdaFunctionCompleted event. It isn't set for other event types.

historyEvent_requestCancelExternalWorkflowExecutionFailedEventAttributes :: Lens' HistoryEvent (Maybe RequestCancelExternalWorkflowExecutionFailedEventAttributes)

If the event is of type RequestCancelExternalWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_timerCanceledEventAttributes :: Lens' HistoryEvent (Maybe TimerCanceledEventAttributes)

If the event is of type TimerCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_workflowExecutionStartedEventAttributes :: Lens' HistoryEvent (Maybe WorkflowExecutionStartedEventAttributes)

If the event is of type WorkflowExecutionStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_activityTaskCompletedEventAttributes :: Lens' HistoryEvent (Maybe ActivityTaskCompletedEventAttributes)

If the event is of type ActivityTaskCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_decisionTaskTimedOutEventAttributes :: Lens' HistoryEvent (Maybe DecisionTaskTimedOutEventAttributes)

If the event is of type DecisionTaskTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_cancelTimerFailedEventAttributes :: Lens' HistoryEvent (Maybe CancelTimerFailedEventAttributes)

If the event is of type CancelTimerFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_childWorkflowExecutionStartedEventAttributes :: Lens' HistoryEvent (Maybe ChildWorkflowExecutionStartedEventAttributes)

If the event is of type ChildWorkflowExecutionStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_activityTaskCanceledEventAttributes :: Lens' HistoryEvent (Maybe ActivityTaskCanceledEventAttributes)

If the event is of type ActivityTaskCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_activityTaskTimedOutEventAttributes :: Lens' HistoryEvent (Maybe ActivityTaskTimedOutEventAttributes)

If the event is of type ActivityTaskTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_decisionTaskStartedEventAttributes :: Lens' HistoryEvent (Maybe DecisionTaskStartedEventAttributes)

If the event is of type DecisionTaskStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_workflowExecutionTerminatedEventAttributes :: Lens' HistoryEvent (Maybe WorkflowExecutionTerminatedEventAttributes)

If the event is of type WorkflowExecutionTerminated then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_childWorkflowExecutionCanceledEventAttributes :: Lens' HistoryEvent (Maybe ChildWorkflowExecutionCanceledEventAttributes)

If the event is of type ChildWorkflowExecutionCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_requestCancelActivityTaskFailedEventAttributes :: Lens' HistoryEvent (Maybe RequestCancelActivityTaskFailedEventAttributes)

If the event is of type RequestCancelActivityTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_lambdaFunctionScheduledEventAttributes :: Lens' HistoryEvent (Maybe LambdaFunctionScheduledEventAttributes)

Provides the details of the LambdaFunctionScheduled event. It isn't set for other event types.

historyEvent_childWorkflowExecutionTimedOutEventAttributes :: Lens' HistoryEvent (Maybe ChildWorkflowExecutionTimedOutEventAttributes)

If the event is of type ChildWorkflowExecutionTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_cancelWorkflowExecutionFailedEventAttributes :: Lens' HistoryEvent (Maybe CancelWorkflowExecutionFailedEventAttributes)

If the event is of type CancelWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_startChildWorkflowExecutionInitiatedEventAttributes :: Lens' HistoryEvent (Maybe StartChildWorkflowExecutionInitiatedEventAttributes)

If the event is of type StartChildWorkflowExecutionInitiated then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_signalExternalWorkflowExecutionFailedEventAttributes :: Lens' HistoryEvent (Maybe SignalExternalWorkflowExecutionFailedEventAttributes)

If the event is of type SignalExternalWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_activityTaskStartedEventAttributes :: Lens' HistoryEvent (Maybe ActivityTaskStartedEventAttributes)

If the event is of type ActivityTaskStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_startLambdaFunctionFailedEventAttributes :: Lens' HistoryEvent (Maybe StartLambdaFunctionFailedEventAttributes)

Provides the details of the StartLambdaFunctionFailed event. It isn't set for other event types.

historyEvent_childWorkflowExecutionTerminatedEventAttributes :: Lens' HistoryEvent (Maybe ChildWorkflowExecutionTerminatedEventAttributes)

If the event is of type ChildWorkflowExecutionTerminated then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_lambdaFunctionFailedEventAttributes :: Lens' HistoryEvent (Maybe LambdaFunctionFailedEventAttributes)

Provides the details of the LambdaFunctionFailed event. It isn't set for other event types.

historyEvent_workflowExecutionCanceledEventAttributes :: Lens' HistoryEvent (Maybe WorkflowExecutionCanceledEventAttributes)

If the event is of type WorkflowExecutionCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_timerStartedEventAttributes :: Lens' HistoryEvent (Maybe TimerStartedEventAttributes)

If the event is of type TimerStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_activityTaskCancelRequestedEventAttributes :: Lens' HistoryEvent (Maybe ActivityTaskCancelRequestedEventAttributes)

If the event is of type ActivityTaskcancelRequested then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_workflowExecutionTimedOutEventAttributes :: Lens' HistoryEvent (Maybe WorkflowExecutionTimedOutEventAttributes)

If the event is of type WorkflowExecutionTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_workflowExecutionSignaledEventAttributes :: Lens' HistoryEvent (Maybe WorkflowExecutionSignaledEventAttributes)

If the event is of type WorkflowExecutionSignaled then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_timerFiredEventAttributes :: Lens' HistoryEvent (Maybe TimerFiredEventAttributes)

If the event is of type TimerFired then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_activityTaskFailedEventAttributes :: Lens' HistoryEvent (Maybe ActivityTaskFailedEventAttributes)

If the event is of type ActivityTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_externalWorkflowExecutionSignaledEventAttributes :: Lens' HistoryEvent (Maybe ExternalWorkflowExecutionSignaledEventAttributes)

If the event is of type ExternalWorkflowExecutionSignaled then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_decisionTaskCompletedEventAttributes :: Lens' HistoryEvent (Maybe DecisionTaskCompletedEventAttributes)

If the event is of type DecisionTaskCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_startChildWorkflowExecutionFailedEventAttributes :: Lens' HistoryEvent (Maybe StartChildWorkflowExecutionFailedEventAttributes)

If the event is of type StartChildWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_childWorkflowExecutionFailedEventAttributes :: Lens' HistoryEvent (Maybe ChildWorkflowExecutionFailedEventAttributes)

If the event is of type ChildWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_failWorkflowExecutionFailedEventAttributes :: Lens' HistoryEvent (Maybe FailWorkflowExecutionFailedEventAttributes)

If the event is of type FailWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_continueAsNewWorkflowExecutionFailedEventAttributes :: Lens' HistoryEvent (Maybe ContinueAsNewWorkflowExecutionFailedEventAttributes)

If the event is of type ContinueAsNewWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_signalExternalWorkflowExecutionInitiatedEventAttributes :: Lens' HistoryEvent (Maybe SignalExternalWorkflowExecutionInitiatedEventAttributes)

If the event is of type SignalExternalWorkflowExecutionInitiated then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_lambdaFunctionTimedOutEventAttributes :: Lens' HistoryEvent (Maybe LambdaFunctionTimedOutEventAttributes)

Provides the details of the LambdaFunctionTimedOut event. It isn't set for other event types.

historyEvent_workflowExecutionFailedEventAttributes :: Lens' HistoryEvent (Maybe WorkflowExecutionFailedEventAttributes)

If the event is of type WorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_workflowExecutionContinuedAsNewEventAttributes :: Lens' HistoryEvent (Maybe WorkflowExecutionContinuedAsNewEventAttributes)

If the event is of type WorkflowExecutionContinuedAsNew then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_externalWorkflowExecutionCancelRequestedEventAttributes :: Lens' HistoryEvent (Maybe ExternalWorkflowExecutionCancelRequestedEventAttributes)

If the event is of type ExternalWorkflowExecutionCancelRequested then this member is set and provides detailed information about the event. It isn't set for other event types.

historyEvent_eventTimestamp :: Lens' HistoryEvent UTCTime

The date and time when the event occurred.

historyEvent_eventType :: Lens' HistoryEvent EventType

The type of the history event.

historyEvent_eventId :: Lens' HistoryEvent Integer

The system generated ID of the event. This ID uniquely identifies the event with in the workflow execution history.