Network.AWS.SWF.Types.LambdaFunctionScheduledEventAttributes

Description

Documentation

data LambdaFunctionScheduledEventAttributes

Provides the details of the LambdaFunctionScheduled event. It isn't set for other event types.

See: newLambdaFunctionScheduledEventAttributes smart constructor.

Constructors

LambdaFunctionScheduledEventAttributes'

Fields

input :: Maybe Text

The input provided to the Lambda task.

control :: Maybe Text

Data attached to the event that the decider can use in subsequent workflow tasks. This data isn't sent to the Lambda task.

startToCloseTimeout :: Maybe Text

The maximum amount of time a worker can take to process the Lambda task.

id :: Text

The unique ID of the Lambda task.

name :: Text

The name of the Lambda function.

decisionTaskCompletedEventId :: Integer

The ID of the LambdaFunctionCompleted event corresponding to the decision that resulted in scheduling this activity task. To help diagnose issues, use this information to trace back the chain of events leading up to this event.

newLambdaFunctionScheduledEventAttributes

Arguments

:: Text
$sel:id:LambdaFunctionScheduledEventAttributes'

-> Text
$sel:name:LambdaFunctionScheduledEventAttributes'

-> Integer
$sel:decisionTaskCompletedEventId:LambdaFunctionScheduledEventAttributes'

-> LambdaFunctionScheduledEventAttributes

Create a value of LambdaFunctionScheduledEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:input:LambdaFunctionScheduledEventAttributes', lambdaFunctionScheduledEventAttributes_input - The input provided to the Lambda task.

$sel:control:LambdaFunctionScheduledEventAttributes', lambdaFunctionScheduledEventAttributes_control - Data attached to the event that the decider can use in subsequent workflow tasks. This data isn't sent to the Lambda task.

$sel:startToCloseTimeout:LambdaFunctionScheduledEventAttributes', lambdaFunctionScheduledEventAttributes_startToCloseTimeout - The maximum amount of time a worker can take to process the Lambda task.

$sel:id:LambdaFunctionScheduledEventAttributes', lambdaFunctionScheduledEventAttributes_id - The unique ID of the Lambda task.

$sel:name:LambdaFunctionScheduledEventAttributes', lambdaFunctionScheduledEventAttributes_name - The name of the Lambda function.

$sel:decisionTaskCompletedEventId:LambdaFunctionScheduledEventAttributes', lambdaFunctionScheduledEventAttributes_decisionTaskCompletedEventId - The ID of the LambdaFunctionCompleted event corresponding to the decision that resulted in scheduling this activity task. To help diagnose issues, use this information to trace back the chain of events leading up to this event.

lambdaFunctionScheduledEventAttributes_input :: Lens' LambdaFunctionScheduledEventAttributes (Maybe Text)

The input provided to the Lambda task.

lambdaFunctionScheduledEventAttributes_control :: Lens' LambdaFunctionScheduledEventAttributes (Maybe Text)

Data attached to the event that the decider can use in subsequent workflow tasks. This data isn't sent to the Lambda task.

lambdaFunctionScheduledEventAttributes_startToCloseTimeout :: Lens' LambdaFunctionScheduledEventAttributes (Maybe Text)

The maximum amount of time a worker can take to process the Lambda task.

lambdaFunctionScheduledEventAttributes_id :: Lens' LambdaFunctionScheduledEventAttributes Text

The unique ID of the Lambda task.

lambdaFunctionScheduledEventAttributes_name :: Lens' LambdaFunctionScheduledEventAttributes Text

The name of the Lambda function.

lambdaFunctionScheduledEventAttributes_decisionTaskCompletedEventId :: Lens' LambdaFunctionScheduledEventAttributes Integer

The ID of the LambdaFunctionCompleted event corresponding to the decision that resulted in scheduling this activity task. To help diagnose issues, use this information to trace back the chain of events leading up to this event.