Network.AWS.SWF.Types.LambdaFunctionTimedOutEventAttributes

Description

Documentation

data LambdaFunctionTimedOutEventAttributes

Provides details of the LambdaFunctionTimedOut event.

See: newLambdaFunctionTimedOutEventAttributes smart constructor.

Constructors

LambdaFunctionTimedOutEventAttributes'

Fields

timeoutType :: Maybe LambdaFunctionTimeoutType

The type of the timeout that caused this event.

scheduledEventId :: Integer

The ID of the LambdaFunctionScheduled event that was recorded when this activity task was scheduled. To help diagnose issues, use this information to trace back the chain of events leading up to this event.

startedEventId :: Integer

The ID of the ActivityTaskStarted event that was recorded when this activity task started. To help diagnose issues, use this information to trace back the chain of events leading up to this event.

newLambdaFunctionTimedOutEventAttributes

Arguments

:: Integer
$sel:scheduledEventId:LambdaFunctionTimedOutEventAttributes'

-> Integer
$sel:startedEventId:LambdaFunctionTimedOutEventAttributes'

-> LambdaFunctionTimedOutEventAttributes

Create a value of LambdaFunctionTimedOutEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeoutType:LambdaFunctionTimedOutEventAttributes', lambdaFunctionTimedOutEventAttributes_timeoutType - The type of the timeout that caused this event.

$sel:scheduledEventId:LambdaFunctionTimedOutEventAttributes', lambdaFunctionTimedOutEventAttributes_scheduledEventId - The ID of the LambdaFunctionScheduled event that was recorded when this activity task was scheduled. To help diagnose issues, use this information to trace back the chain of events leading up to this event.

$sel:startedEventId:LambdaFunctionTimedOutEventAttributes', lambdaFunctionTimedOutEventAttributes_startedEventId - The ID of the ActivityTaskStarted event that was recorded when this activity task started. To help diagnose issues, use this information to trace back the chain of events leading up to this event.

lambdaFunctionTimedOutEventAttributes_timeoutType :: Lens' LambdaFunctionTimedOutEventAttributes (Maybe LambdaFunctionTimeoutType)

The type of the timeout that caused this event.

lambdaFunctionTimedOutEventAttributes_scheduledEventId :: Lens' LambdaFunctionTimedOutEventAttributes Integer

The ID of the LambdaFunctionScheduled event that was recorded when this activity task was scheduled. To help diagnose issues, use this information to trace back the chain of events leading up to this event.

lambdaFunctionTimedOutEventAttributes_startedEventId :: Lens' LambdaFunctionTimedOutEventAttributes Integer

The ID of the ActivityTaskStarted event that was recorded when this activity task started. To help diagnose issues, use this information to trace back the chain of events leading up to this event.