Network.AWS.SWF.Types.ScheduleActivityTaskDecisionAttributes

Description

Documentation

data ScheduleActivityTaskDecisionAttributes

Provides the details of the ScheduleActivityTask decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

activityType.name – String constraint. The key is swf:activityType.name.

activityType.version – String constraint. The key is swf:activityType.version.

taskList – String constraint. The key is swf:taskList.name.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

See: newScheduleActivityTaskDecisionAttributes smart constructor.

Constructors

ScheduleActivityTaskDecisionAttributes'

Fields

input :: Maybe Text

The input provided to the activity task.

heartbeatTimeout :: Maybe Text

If set, specifies the maximum time before which a worker processing a task of this type must report progress by calling RecordActivityTaskHeartbeat. If the timeout is exceeded, the activity task is automatically timed out. If the worker subsequently attempts to record a heartbeat or returns a result, it is ignored. This overrides the default heartbeat timeout specified when registering the activity type using RegisterActivityType.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

scheduleToCloseTimeout :: Maybe Text

The maximum duration for this activity task.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

A schedule-to-close timeout for this activity task must be specified either as a default for the activity type or through this field. If neither this field is set nor a default schedule-to-close timeout was specified at registration time then a fault is returned.

scheduleToStartTimeout :: Maybe Text

If set, specifies the maximum duration the activity task can wait to be assigned to a worker. This overrides the default schedule-to-start timeout specified when registering the activity type using RegisterActivityType.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

A schedule-to-start timeout for this activity task must be specified either as a default for the activity type or through this field. If neither this field is set nor a default schedule-to-start timeout was specified at registration time then a fault is returned.

taskPriority :: Maybe Text

If set, specifies the priority with which the activity task is to be assigned to a worker. This overrides the defaultTaskPriority specified when registering the activity type using RegisterActivityType. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

taskList :: Maybe TaskList

If set, specifies the name of the task list in which to schedule the activity task. If not specified, the defaultTaskList registered with the activity type is used.

A task list for this activity task must be specified either as a default for the activity type or through this field. If neither this field is set nor a default task list was specified at registration time then a fault is returned.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not contain the literal string arn.

control :: Maybe Text

Data attached to the event that can be used by the decider in subsequent workflow tasks. This data isn't sent to the activity.

startToCloseTimeout :: Maybe Text

If set, specifies the maximum duration a worker may take to process this activity task. This overrides the default start-to-close timeout specified when registering the activity type using RegisterActivityType.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

A start-to-close timeout for this activity task must be specified either as a default for the activity type or through this field. If neither this field is set nor a default start-to-close timeout was specified at registration time then a fault is returned.

activityType :: ActivityType

The type of the activity task to schedule.

activityId :: Text

The activityId of the activity task.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not contain the literal string arn.

newScheduleActivityTaskDecisionAttributes

Arguments

:: ActivityType
$sel:activityType:ScheduleActivityTaskDecisionAttributes'

-> Text
$sel:activityId:ScheduleActivityTaskDecisionAttributes'

-> ScheduleActivityTaskDecisionAttributes

Create a value of ScheduleActivityTaskDecisionAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:input:ScheduleActivityTaskDecisionAttributes', scheduleActivityTaskDecisionAttributes_input - The input provided to the activity task.

$sel:heartbeatTimeout:ScheduleActivityTaskDecisionAttributes', scheduleActivityTaskDecisionAttributes_heartbeatTimeout - If set, specifies the maximum time before which a worker processing a task of this type must report progress by calling RecordActivityTaskHeartbeat. If the timeout is exceeded, the activity task is automatically timed out. If the worker subsequently attempts to record a heartbeat or returns a result, it is ignored. This overrides the default heartbeat timeout specified when registering the activity type using RegisterActivityType.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

$sel:scheduleToCloseTimeout:ScheduleActivityTaskDecisionAttributes', scheduleActivityTaskDecisionAttributes_scheduleToCloseTimeout - The maximum duration for this activity task.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

A schedule-to-close timeout for this activity task must be specified either as a default for the activity type or through this field. If neither this field is set nor a default schedule-to-close timeout was specified at registration time then a fault is returned.

$sel:scheduleToStartTimeout:ScheduleActivityTaskDecisionAttributes', scheduleActivityTaskDecisionAttributes_scheduleToStartTimeout - If set, specifies the maximum duration the activity task can wait to be assigned to a worker. This overrides the default schedule-to-start timeout specified when registering the activity type using RegisterActivityType.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

A schedule-to-start timeout for this activity task must be specified either as a default for the activity type or through this field. If neither this field is set nor a default schedule-to-start timeout was specified at registration time then a fault is returned.

$sel:taskPriority:ScheduleActivityTaskDecisionAttributes', scheduleActivityTaskDecisionAttributes_taskPriority - If set, specifies the priority with which the activity task is to be assigned to a worker. This overrides the defaultTaskPriority specified when registering the activity type using RegisterActivityType. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

$sel:taskList:ScheduleActivityTaskDecisionAttributes', scheduleActivityTaskDecisionAttributes_taskList - If set, specifies the name of the task list in which to schedule the activity task. If not specified, the defaultTaskList registered with the activity type is used.

A task list for this activity task must be specified either as a default for the activity type or through this field. If neither this field is set nor a default task list was specified at registration time then a fault is returned.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not contain the literal string arn.

$sel:control:ScheduleActivityTaskDecisionAttributes', scheduleActivityTaskDecisionAttributes_control - Data attached to the event that can be used by the decider in subsequent workflow tasks. This data isn't sent to the activity.

$sel:startToCloseTimeout:ScheduleActivityTaskDecisionAttributes', scheduleActivityTaskDecisionAttributes_startToCloseTimeout - If set, specifies the maximum duration a worker may take to process this activity task. This overrides the default start-to-close timeout specified when registering the activity type using RegisterActivityType.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

A start-to-close timeout for this activity task must be specified either as a default for the activity type or through this field. If neither this field is set nor a default start-to-close timeout was specified at registration time then a fault is returned.

$sel:activityType:ScheduleActivityTaskDecisionAttributes', scheduleActivityTaskDecisionAttributes_activityType - The type of the activity task to schedule.

$sel:activityId:ScheduleActivityTaskDecisionAttributes', scheduleActivityTaskDecisionAttributes_activityId - The activityId of the activity task.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not contain the literal string arn.

scheduleActivityTaskDecisionAttributes_input :: Lens' ScheduleActivityTaskDecisionAttributes (Maybe Text)

The input provided to the activity task.

scheduleActivityTaskDecisionAttributes_heartbeatTimeout :: Lens' ScheduleActivityTaskDecisionAttributes (Maybe Text)

If set, specifies the maximum time before which a worker processing a task of this type must report progress by calling RecordActivityTaskHeartbeat. If the timeout is exceeded, the activity task is automatically timed out. If the worker subsequently attempts to record a heartbeat or returns a result, it is ignored. This overrides the default heartbeat timeout specified when registering the activity type using RegisterActivityType.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

scheduleActivityTaskDecisionAttributes_scheduleToCloseTimeout :: Lens' ScheduleActivityTaskDecisionAttributes (Maybe Text)

The maximum duration for this activity task.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

A schedule-to-close timeout for this activity task must be specified either as a default for the activity type or through this field. If neither this field is set nor a default schedule-to-close timeout was specified at registration time then a fault is returned.

scheduleActivityTaskDecisionAttributes_scheduleToStartTimeout :: Lens' ScheduleActivityTaskDecisionAttributes (Maybe Text)

If set, specifies the maximum duration the activity task can wait to be assigned to a worker. This overrides the default schedule-to-start timeout specified when registering the activity type using RegisterActivityType.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

A schedule-to-start timeout for this activity task must be specified either as a default for the activity type or through this field. If neither this field is set nor a default schedule-to-start timeout was specified at registration time then a fault is returned.

scheduleActivityTaskDecisionAttributes_taskPriority :: Lens' ScheduleActivityTaskDecisionAttributes (Maybe Text)

If set, specifies the priority with which the activity task is to be assigned to a worker. This overrides the defaultTaskPriority specified when registering the activity type using RegisterActivityType. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

scheduleActivityTaskDecisionAttributes_taskList :: Lens' ScheduleActivityTaskDecisionAttributes (Maybe TaskList)

If set, specifies the name of the task list in which to schedule the activity task. If not specified, the defaultTaskList registered with the activity type is used.

A task list for this activity task must be specified either as a default for the activity type or through this field. If neither this field is set nor a default task list was specified at registration time then a fault is returned.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not contain the literal string arn.

scheduleActivityTaskDecisionAttributes_control :: Lens' ScheduleActivityTaskDecisionAttributes (Maybe Text)

Data attached to the event that can be used by the decider in subsequent workflow tasks. This data isn't sent to the activity.

scheduleActivityTaskDecisionAttributes_startToCloseTimeout :: Lens' ScheduleActivityTaskDecisionAttributes (Maybe Text)

If set, specifies the maximum duration a worker may take to process this activity task. This overrides the default start-to-close timeout specified when registering the activity type using RegisterActivityType.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

A start-to-close timeout for this activity task must be specified either as a default for the activity type or through this field. If neither this field is set nor a default start-to-close timeout was specified at registration time then a fault is returned.

scheduleActivityTaskDecisionAttributes_activityType :: Lens' ScheduleActivityTaskDecisionAttributes ActivityType

The type of the activity task to schedule.

scheduleActivityTaskDecisionAttributes_activityId :: Lens' ScheduleActivityTaskDecisionAttributes Text

The activityId of the activity task.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not contain the literal string arn.