Network.AWS.SWF.Types.ScheduleLambdaFunctionDecisionAttributes

Description

Documentation

data ScheduleLambdaFunctionDecisionAttributes

Decision attributes specified in scheduleLambdaFunctionDecisionAttributes within the list of decisions decisions passed to RespondDecisionTaskCompleted.

See: newScheduleLambdaFunctionDecisionAttributes smart constructor.

Constructors

ScheduleLambdaFunctionDecisionAttributes'

Fields

input :: Maybe Text

The optional input data to be supplied to the Lambda function.

control :: Maybe Text

The data attached to the event that the decider can use in subsequent workflow tasks. This data isn't sent to the Lambda task.

startToCloseTimeout :: Maybe Text

The timeout value, in seconds, after which the Lambda function is considered to be failed once it has started. This can be any integer from 1-300 (1s-5m). If no value is supplied, than a default value of 300s is assumed.

id :: Text

A string that identifies the Lambda function execution in the event history.

name :: Text

The name, or ARN, of the Lambda function to schedule.

newScheduleLambdaFunctionDecisionAttributes

Arguments

:: Text
$sel:id:ScheduleLambdaFunctionDecisionAttributes'

-> Text
$sel:name:ScheduleLambdaFunctionDecisionAttributes'

-> ScheduleLambdaFunctionDecisionAttributes

Create a value of ScheduleLambdaFunctionDecisionAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:input:ScheduleLambdaFunctionDecisionAttributes', scheduleLambdaFunctionDecisionAttributes_input - The optional input data to be supplied to the Lambda function.

$sel:control:ScheduleLambdaFunctionDecisionAttributes', scheduleLambdaFunctionDecisionAttributes_control - The data attached to the event that the decider can use in subsequent workflow tasks. This data isn't sent to the Lambda task.

$sel:startToCloseTimeout:ScheduleLambdaFunctionDecisionAttributes', scheduleLambdaFunctionDecisionAttributes_startToCloseTimeout - The timeout value, in seconds, after which the Lambda function is considered to be failed once it has started. This can be any integer from 1-300 (1s-5m). If no value is supplied, than a default value of 300s is assumed.

$sel:id:ScheduleLambdaFunctionDecisionAttributes', scheduleLambdaFunctionDecisionAttributes_id - A string that identifies the Lambda function execution in the event history.

$sel:name:ScheduleLambdaFunctionDecisionAttributes', scheduleLambdaFunctionDecisionAttributes_name - The name, or ARN, of the Lambda function to schedule.

scheduleLambdaFunctionDecisionAttributes_input :: Lens' ScheduleLambdaFunctionDecisionAttributes (Maybe Text)

The optional input data to be supplied to the Lambda function.

scheduleLambdaFunctionDecisionAttributes_control :: Lens' ScheduleLambdaFunctionDecisionAttributes (Maybe Text)

The data attached to the event that the decider can use in subsequent workflow tasks. This data isn't sent to the Lambda task.

scheduleLambdaFunctionDecisionAttributes_startToCloseTimeout :: Lens' ScheduleLambdaFunctionDecisionAttributes (Maybe Text)

The timeout value, in seconds, after which the Lambda function is considered to be failed once it has started. This can be any integer from 1-300 (1s-5m). If no value is supplied, than a default value of 300s is assumed.

scheduleLambdaFunctionDecisionAttributes_id :: Lens' ScheduleLambdaFunctionDecisionAttributes Text

A string that identifies the Lambda function execution in the event history.

scheduleLambdaFunctionDecisionAttributes_name :: Lens' ScheduleLambdaFunctionDecisionAttributes Text

The name, or ARN, of the Lambda function to schedule.