Network.AWS.SWF.Types.SignalExternalWorkflowExecutionDecisionAttributes

Description

Documentation

data SignalExternalWorkflowExecutionDecisionAttributes

Provides the details of the SignalExternalWorkflowExecution decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

See: newSignalExternalWorkflowExecutionDecisionAttributes smart constructor.

Constructors

SignalExternalWorkflowExecutionDecisionAttributes'

Fields

runId :: Maybe Text

The runId of the workflow execution to be signaled.

input :: Maybe Text

The input data to be provided with the signal. The target workflow execution uses the signal name and input data to process the signal.

control :: Maybe Text

The data attached to the event that can be used by the decider in subsequent decision tasks.

workflowId :: Text

The workflowId of the workflow execution to be signaled.

signalName :: Text

The name of the signal.The target workflow execution uses the signal name and input to process the signal.

newSignalExternalWorkflowExecutionDecisionAttributes

Arguments

:: Text
$sel:workflowId:SignalExternalWorkflowExecutionDecisionAttributes'

-> Text
$sel:signalName:SignalExternalWorkflowExecutionDecisionAttributes'

-> SignalExternalWorkflowExecutionDecisionAttributes

Create a value of SignalExternalWorkflowExecutionDecisionAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:runId:SignalExternalWorkflowExecutionDecisionAttributes', signalExternalWorkflowExecutionDecisionAttributes_runId - The runId of the workflow execution to be signaled.

$sel:input:SignalExternalWorkflowExecutionDecisionAttributes', signalExternalWorkflowExecutionDecisionAttributes_input - The input data to be provided with the signal. The target workflow execution uses the signal name and input data to process the signal.

$sel:control:SignalExternalWorkflowExecutionDecisionAttributes', signalExternalWorkflowExecutionDecisionAttributes_control - The data attached to the event that can be used by the decider in subsequent decision tasks.

$sel:workflowId:SignalExternalWorkflowExecutionDecisionAttributes', signalExternalWorkflowExecutionDecisionAttributes_workflowId - The workflowId of the workflow execution to be signaled.

$sel:signalName:SignalExternalWorkflowExecutionDecisionAttributes', signalExternalWorkflowExecutionDecisionAttributes_signalName - The name of the signal.The target workflow execution uses the signal name and input to process the signal.

signalExternalWorkflowExecutionDecisionAttributes_runId :: Lens' SignalExternalWorkflowExecutionDecisionAttributes (Maybe Text)

The runId of the workflow execution to be signaled.

signalExternalWorkflowExecutionDecisionAttributes_input :: Lens' SignalExternalWorkflowExecutionDecisionAttributes (Maybe Text)

The input data to be provided with the signal. The target workflow execution uses the signal name and input data to process the signal.

signalExternalWorkflowExecutionDecisionAttributes_control :: Lens' SignalExternalWorkflowExecutionDecisionAttributes (Maybe Text)

The data attached to the event that can be used by the decider in subsequent decision tasks.

signalExternalWorkflowExecutionDecisionAttributes_workflowId :: Lens' SignalExternalWorkflowExecutionDecisionAttributes Text

The workflowId of the workflow execution to be signaled.

signalExternalWorkflowExecutionDecisionAttributes_signalName :: Lens' SignalExternalWorkflowExecutionDecisionAttributes Text

The name of the signal.The target workflow execution uses the signal name and input to process the signal.