Network.AWS.SWF.Types.SignalExternalWorkflowExecutionInitiatedEventAttributes

Description

Documentation

data SignalExternalWorkflowExecutionInitiatedEventAttributes

Provides the details of the SignalExternalWorkflowExecutionInitiated event.

See: newSignalExternalWorkflowExecutionInitiatedEventAttributes smart constructor.

Constructors

SignalExternalWorkflowExecutionInitiatedEventAttributes'

Fields

runId :: Maybe Text

The runId of the external workflow execution to send the signal to.

input :: Maybe Text

The input provided to the signal.

control :: Maybe Text

Data attached to the event that can be used by the decider in subsequent decision tasks.

workflowId :: Text

The workflowId of the external workflow execution.

signalName :: Text

The name of the signal.

decisionTaskCompletedEventId :: Integer

The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the SignalExternalWorkflowExecution decision for this signal. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

newSignalExternalWorkflowExecutionInitiatedEventAttributes

Arguments

:: Text
$sel:workflowId:SignalExternalWorkflowExecutionInitiatedEventAttributes'

-> Text
$sel:signalName:SignalExternalWorkflowExecutionInitiatedEventAttributes'

-> Integer
$sel:decisionTaskCompletedEventId:SignalExternalWorkflowExecutionInitiatedEventAttributes'

-> SignalExternalWorkflowExecutionInitiatedEventAttributes

Create a value of SignalExternalWorkflowExecutionInitiatedEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:runId:SignalExternalWorkflowExecutionInitiatedEventAttributes', signalExternalWorkflowExecutionInitiatedEventAttributes_runId - The runId of the external workflow execution to send the signal to.

$sel:input:SignalExternalWorkflowExecutionInitiatedEventAttributes', signalExternalWorkflowExecutionInitiatedEventAttributes_input - The input provided to the signal.

$sel:control:SignalExternalWorkflowExecutionInitiatedEventAttributes', signalExternalWorkflowExecutionInitiatedEventAttributes_control - Data attached to the event that can be used by the decider in subsequent decision tasks.

$sel:workflowId:SignalExternalWorkflowExecutionInitiatedEventAttributes', signalExternalWorkflowExecutionInitiatedEventAttributes_workflowId - The workflowId of the external workflow execution.

$sel:signalName:SignalExternalWorkflowExecutionInitiatedEventAttributes', signalExternalWorkflowExecutionInitiatedEventAttributes_signalName - The name of the signal.

$sel:decisionTaskCompletedEventId:SignalExternalWorkflowExecutionInitiatedEventAttributes', signalExternalWorkflowExecutionInitiatedEventAttributes_decisionTaskCompletedEventId - The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the SignalExternalWorkflowExecution decision for this signal. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

signalExternalWorkflowExecutionInitiatedEventAttributes_runId :: Lens' SignalExternalWorkflowExecutionInitiatedEventAttributes (Maybe Text)

The runId of the external workflow execution to send the signal to.

signalExternalWorkflowExecutionInitiatedEventAttributes_input :: Lens' SignalExternalWorkflowExecutionInitiatedEventAttributes (Maybe Text)

The input provided to the signal.

signalExternalWorkflowExecutionInitiatedEventAttributes_control :: Lens' SignalExternalWorkflowExecutionInitiatedEventAttributes (Maybe Text)

Data attached to the event that can be used by the decider in subsequent decision tasks.

signalExternalWorkflowExecutionInitiatedEventAttributes_workflowId :: Lens' SignalExternalWorkflowExecutionInitiatedEventAttributes Text

The workflowId of the external workflow execution.

signalExternalWorkflowExecutionInitiatedEventAttributes_signalName :: Lens' SignalExternalWorkflowExecutionInitiatedEventAttributes Text

The name of the signal.

signalExternalWorkflowExecutionInitiatedEventAttributes_decisionTaskCompletedEventId :: Lens' SignalExternalWorkflowExecutionInitiatedEventAttributes Integer

The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the SignalExternalWorkflowExecution decision for this signal. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.