Network.AWS.SWF.Types.StartChildWorkflowExecutionDecisionAttributes

Description

Documentation

data StartChildWorkflowExecutionDecisionAttributes

Provides the details of the StartChildWorkflowExecution decision.

Access Control

You can use IAM policies to control this decision's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

tagList.member.N – The key is "swf:tagList.N" where N is the tag number from 0 to 4, inclusive.

taskList – String constraint. The key is swf:taskList.name.

workflowType.name – String constraint. The key is swf:workflowType.name.

workflowType.version – String constraint. The key is swf:workflowType.version.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

See: newStartChildWorkflowExecutionDecisionAttributes smart constructor.

Constructors

StartChildWorkflowExecutionDecisionAttributes'

Fields

control :: Maybe Text

The data attached to the event that can be used by the decider in subsequent workflow tasks. This data isn't sent to the child workflow execution.

tagList :: Maybe [Text]

The list of tags to associate with the child workflow execution. A maximum of 5 tags can be specified. You can list workflow executions with a specific tag by calling ListOpenWorkflowExecutions or ListClosedWorkflowExecutions and specifying a TagFilter.

taskStartToCloseTimeout :: Maybe Text

Specifies the maximum duration of decision tasks for this workflow execution. This parameter overrides the defaultTaskStartToCloseTimout specified when registering the workflow type using RegisterWorkflowType.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

A task start-to-close timeout for this workflow execution must be specified either as a default for the workflow type or through this parameter. If neither this parameter is set nor a default task start-to-close timeout was specified at registration time then a fault is returned.

lambdaRole :: Maybe Text

The IAM role attached to the child workflow execution.

input :: Maybe Text

The input to be provided to the workflow execution.

executionStartToCloseTimeout :: Maybe Text

The total duration for this workflow execution. This overrides the defaultExecutionStartToCloseTimeout specified when registering the workflow type.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

An execution start-to-close timeout for this workflow execution must be specified either as a default for the workflow type or through this parameter. If neither this parameter is set nor a default execution start-to-close timeout was specified at registration time then a fault is returned.

taskList :: Maybe TaskList

The name of the task list to be used for decision tasks of the child workflow execution.

A task list for this workflow execution must be specified either as a default for the workflow type or through this parameter. If neither this parameter is set nor a default task list was specified at registration time then a fault is returned.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not contain the literal string arn.

taskPriority :: Maybe Text

A task priority that, if set, specifies the priority for a decision task of this workflow execution. This overrides the defaultTaskPriority specified when registering the workflow type. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

childPolicy :: Maybe ChildPolicy

If set, specifies the policy to use for the child workflow executions if the workflow execution being started is terminated by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout. This policy overrides the default child policy specified when registering the workflow type using RegisterWorkflowType.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

A child policy for this workflow execution must be specified either as a default for the workflow type or through this parameter. If neither this parameter is set nor a default child policy was specified at registration time then a fault is returned.

workflowType :: WorkflowType

The type of the workflow execution to be started.

workflowId :: Text

The workflowId of the workflow execution.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not contain the literal string arn.

newStartChildWorkflowExecutionDecisionAttributes

Arguments

:: WorkflowType
$sel:workflowType:StartChildWorkflowExecutionDecisionAttributes'

-> Text
$sel:workflowId:StartChildWorkflowExecutionDecisionAttributes'

-> StartChildWorkflowExecutionDecisionAttributes

Create a value of StartChildWorkflowExecutionDecisionAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:control:StartChildWorkflowExecutionDecisionAttributes', startChildWorkflowExecutionDecisionAttributes_control - The data attached to the event that can be used by the decider in subsequent workflow tasks. This data isn't sent to the child workflow execution.

$sel:tagList:StartChildWorkflowExecutionDecisionAttributes', startChildWorkflowExecutionDecisionAttributes_tagList - The list of tags to associate with the child workflow execution. A maximum of 5 tags can be specified. You can list workflow executions with a specific tag by calling ListOpenWorkflowExecutions or ListClosedWorkflowExecutions and specifying a TagFilter.

$sel:taskStartToCloseTimeout:StartChildWorkflowExecutionDecisionAttributes', startChildWorkflowExecutionDecisionAttributes_taskStartToCloseTimeout - Specifies the maximum duration of decision tasks for this workflow execution. This parameter overrides the defaultTaskStartToCloseTimout specified when registering the workflow type using RegisterWorkflowType.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

A task start-to-close timeout for this workflow execution must be specified either as a default for the workflow type or through this parameter. If neither this parameter is set nor a default task start-to-close timeout was specified at registration time then a fault is returned.

$sel:lambdaRole:StartChildWorkflowExecutionDecisionAttributes', startChildWorkflowExecutionDecisionAttributes_lambdaRole - The IAM role attached to the child workflow execution.

$sel:input:StartChildWorkflowExecutionDecisionAttributes', startChildWorkflowExecutionDecisionAttributes_input - The input to be provided to the workflow execution.

$sel:executionStartToCloseTimeout:StartChildWorkflowExecutionDecisionAttributes', startChildWorkflowExecutionDecisionAttributes_executionStartToCloseTimeout - The total duration for this workflow execution. This overrides the defaultExecutionStartToCloseTimeout specified when registering the workflow type.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

An execution start-to-close timeout for this workflow execution must be specified either as a default for the workflow type or through this parameter. If neither this parameter is set nor a default execution start-to-close timeout was specified at registration time then a fault is returned.

$sel:taskList:StartChildWorkflowExecutionDecisionAttributes', startChildWorkflowExecutionDecisionAttributes_taskList - The name of the task list to be used for decision tasks of the child workflow execution.

A task list for this workflow execution must be specified either as a default for the workflow type or through this parameter. If neither this parameter is set nor a default task list was specified at registration time then a fault is returned.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not contain the literal string arn.

$sel:taskPriority:StartChildWorkflowExecutionDecisionAttributes', startChildWorkflowExecutionDecisionAttributes_taskPriority - A task priority that, if set, specifies the priority for a decision task of this workflow execution. This overrides the defaultTaskPriority specified when registering the workflow type. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

$sel:childPolicy:StartChildWorkflowExecutionDecisionAttributes', startChildWorkflowExecutionDecisionAttributes_childPolicy - If set, specifies the policy to use for the child workflow executions if the workflow execution being started is terminated by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout. This policy overrides the default child policy specified when registering the workflow type using RegisterWorkflowType.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

A child policy for this workflow execution must be specified either as a default for the workflow type or through this parameter. If neither this parameter is set nor a default child policy was specified at registration time then a fault is returned.

$sel:workflowType:StartChildWorkflowExecutionDecisionAttributes', startChildWorkflowExecutionDecisionAttributes_workflowType - The type of the workflow execution to be started.

$sel:workflowId:StartChildWorkflowExecutionDecisionAttributes', startChildWorkflowExecutionDecisionAttributes_workflowId - The workflowId of the workflow execution.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not contain the literal string arn.

startChildWorkflowExecutionDecisionAttributes_control :: Lens' StartChildWorkflowExecutionDecisionAttributes (Maybe Text)

The data attached to the event that can be used by the decider in subsequent workflow tasks. This data isn't sent to the child workflow execution.

startChildWorkflowExecutionDecisionAttributes_tagList :: Lens' StartChildWorkflowExecutionDecisionAttributes (Maybe [Text])

The list of tags to associate with the child workflow execution. A maximum of 5 tags can be specified. You can list workflow executions with a specific tag by calling ListOpenWorkflowExecutions or ListClosedWorkflowExecutions and specifying a TagFilter.

startChildWorkflowExecutionDecisionAttributes_taskStartToCloseTimeout :: Lens' StartChildWorkflowExecutionDecisionAttributes (Maybe Text)

Specifies the maximum duration of decision tasks for this workflow execution. This parameter overrides the defaultTaskStartToCloseTimout specified when registering the workflow type using RegisterWorkflowType.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

A task start-to-close timeout for this workflow execution must be specified either as a default for the workflow type or through this parameter. If neither this parameter is set nor a default task start-to-close timeout was specified at registration time then a fault is returned.

startChildWorkflowExecutionDecisionAttributes_lambdaRole :: Lens' StartChildWorkflowExecutionDecisionAttributes (Maybe Text)

The IAM role attached to the child workflow execution.

startChildWorkflowExecutionDecisionAttributes_input :: Lens' StartChildWorkflowExecutionDecisionAttributes (Maybe Text)

The input to be provided to the workflow execution.

startChildWorkflowExecutionDecisionAttributes_executionStartToCloseTimeout :: Lens' StartChildWorkflowExecutionDecisionAttributes (Maybe Text)

The total duration for this workflow execution. This overrides the defaultExecutionStartToCloseTimeout specified when registering the workflow type.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

An execution start-to-close timeout for this workflow execution must be specified either as a default for the workflow type or through this parameter. If neither this parameter is set nor a default execution start-to-close timeout was specified at registration time then a fault is returned.

startChildWorkflowExecutionDecisionAttributes_taskList :: Lens' StartChildWorkflowExecutionDecisionAttributes (Maybe TaskList)

The name of the task list to be used for decision tasks of the child workflow execution.

A task list for this workflow execution must be specified either as a default for the workflow type or through this parameter. If neither this parameter is set nor a default task list was specified at registration time then a fault is returned.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not contain the literal string arn.

startChildWorkflowExecutionDecisionAttributes_taskPriority :: Lens' StartChildWorkflowExecutionDecisionAttributes (Maybe Text)

A task priority that, if set, specifies the priority for a decision task of this workflow execution. This overrides the defaultTaskPriority specified when registering the workflow type. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

startChildWorkflowExecutionDecisionAttributes_childPolicy :: Lens' StartChildWorkflowExecutionDecisionAttributes (Maybe ChildPolicy)

If set, specifies the policy to use for the child workflow executions if the workflow execution being started is terminated by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout. This policy overrides the default child policy specified when registering the workflow type using RegisterWorkflowType.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

A child policy for this workflow execution must be specified either as a default for the workflow type or through this parameter. If neither this parameter is set nor a default child policy was specified at registration time then a fault is returned.

startChildWorkflowExecutionDecisionAttributes_workflowType :: Lens' StartChildWorkflowExecutionDecisionAttributes WorkflowType

The type of the workflow execution to be started.

startChildWorkflowExecutionDecisionAttributes_workflowId :: Lens' StartChildWorkflowExecutionDecisionAttributes Text

The workflowId of the workflow execution.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not contain the literal string arn.