Network.AWS.SWF.Types.StartChildWorkflowExecutionInitiatedEventAttributes

Description

Documentation

data StartChildWorkflowExecutionInitiatedEventAttributes

Provides the details of the StartChildWorkflowExecutionInitiated event.

See: newStartChildWorkflowExecutionInitiatedEventAttributes smart constructor.

Constructors

StartChildWorkflowExecutionInitiatedEventAttributes'

Fields

input :: Maybe Text

The inputs provided to the child workflow execution.

lambdaRole :: Maybe Text

The IAM role to attach to the child workflow execution.

taskPriority :: Maybe Text

The priority assigned for the decision tasks for this workflow execution. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

control :: Maybe Text

Data attached to the event that can be used by the decider in subsequent decision tasks. This data isn't sent to the activity.

executionStartToCloseTimeout :: Maybe Text

The maximum duration for the child workflow execution. If the workflow execution isn't closed within this duration, it is timed out and force-terminated.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

taskStartToCloseTimeout :: Maybe Text

The maximum duration allowed for the decision tasks for this workflow execution.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

tagList :: Maybe [Text]

The list of tags to associated with the child workflow execution.

workflowId :: Text

The workflowId of the child workflow execution.

workflowType :: WorkflowType

The type of the child workflow execution.

taskList :: TaskList

The name of the task list used for the decision tasks of the child workflow execution.

decisionTaskCompletedEventId :: Integer

The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the StartChildWorkflowExecution Decision to request this child workflow execution. This information can be useful for diagnosing problems by tracing back the cause of events.

childPolicy :: ChildPolicy

The policy to use for the child workflow executions if this execution gets terminated by explicitly calling the TerminateWorkflowExecution action or due to an expired timeout.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

newStartChildWorkflowExecutionInitiatedEventAttributes

Arguments

:: Text
$sel:workflowId:StartChildWorkflowExecutionInitiatedEventAttributes'

-> WorkflowType
$sel:workflowType:StartChildWorkflowExecutionInitiatedEventAttributes'

-> TaskList
$sel:taskList:StartChildWorkflowExecutionInitiatedEventAttributes'

-> Integer
$sel:decisionTaskCompletedEventId:StartChildWorkflowExecutionInitiatedEventAttributes'

-> ChildPolicy
$sel:childPolicy:StartChildWorkflowExecutionInitiatedEventAttributes'

-> StartChildWorkflowExecutionInitiatedEventAttributes

Create a value of StartChildWorkflowExecutionInitiatedEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:input:StartChildWorkflowExecutionInitiatedEventAttributes', startChildWorkflowExecutionInitiatedEventAttributes_input - The inputs provided to the child workflow execution.

$sel:lambdaRole:StartChildWorkflowExecutionInitiatedEventAttributes', startChildWorkflowExecutionInitiatedEventAttributes_lambdaRole - The IAM role to attach to the child workflow execution.

$sel:taskPriority:StartChildWorkflowExecutionInitiatedEventAttributes', startChildWorkflowExecutionInitiatedEventAttributes_taskPriority - The priority assigned for the decision tasks for this workflow execution. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

$sel:control:StartChildWorkflowExecutionInitiatedEventAttributes', startChildWorkflowExecutionInitiatedEventAttributes_control - Data attached to the event that can be used by the decider in subsequent decision tasks. This data isn't sent to the activity.

$sel:executionStartToCloseTimeout:StartChildWorkflowExecutionInitiatedEventAttributes', startChildWorkflowExecutionInitiatedEventAttributes_executionStartToCloseTimeout - The maximum duration for the child workflow execution. If the workflow execution isn't closed within this duration, it is timed out and force-terminated.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

$sel:taskStartToCloseTimeout:StartChildWorkflowExecutionInitiatedEventAttributes', startChildWorkflowExecutionInitiatedEventAttributes_taskStartToCloseTimeout - The maximum duration allowed for the decision tasks for this workflow execution.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

$sel:tagList:StartChildWorkflowExecutionInitiatedEventAttributes', startChildWorkflowExecutionInitiatedEventAttributes_tagList - The list of tags to associated with the child workflow execution.

$sel:workflowId:StartChildWorkflowExecutionInitiatedEventAttributes', startChildWorkflowExecutionInitiatedEventAttributes_workflowId - The workflowId of the child workflow execution.

$sel:workflowType:StartChildWorkflowExecutionInitiatedEventAttributes', startChildWorkflowExecutionInitiatedEventAttributes_workflowType - The type of the child workflow execution.

$sel:taskList:StartChildWorkflowExecutionInitiatedEventAttributes', startChildWorkflowExecutionInitiatedEventAttributes_taskList - The name of the task list used for the decision tasks of the child workflow execution.

$sel:decisionTaskCompletedEventId:StartChildWorkflowExecutionInitiatedEventAttributes', startChildWorkflowExecutionInitiatedEventAttributes_decisionTaskCompletedEventId - The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the StartChildWorkflowExecution Decision to request this child workflow execution. This information can be useful for diagnosing problems by tracing back the cause of events.

$sel:childPolicy:StartChildWorkflowExecutionInitiatedEventAttributes', startChildWorkflowExecutionInitiatedEventAttributes_childPolicy - The policy to use for the child workflow executions if this execution gets terminated by explicitly calling the TerminateWorkflowExecution action or due to an expired timeout.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

startChildWorkflowExecutionInitiatedEventAttributes_input :: Lens' StartChildWorkflowExecutionInitiatedEventAttributes (Maybe Text)

The inputs provided to the child workflow execution.

startChildWorkflowExecutionInitiatedEventAttributes_lambdaRole :: Lens' StartChildWorkflowExecutionInitiatedEventAttributes (Maybe Text)

The IAM role to attach to the child workflow execution.

startChildWorkflowExecutionInitiatedEventAttributes_taskPriority :: Lens' StartChildWorkflowExecutionInitiatedEventAttributes (Maybe Text)

The priority assigned for the decision tasks for this workflow execution. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

startChildWorkflowExecutionInitiatedEventAttributes_control :: Lens' StartChildWorkflowExecutionInitiatedEventAttributes (Maybe Text)

Data attached to the event that can be used by the decider in subsequent decision tasks. This data isn't sent to the activity.

startChildWorkflowExecutionInitiatedEventAttributes_executionStartToCloseTimeout :: Lens' StartChildWorkflowExecutionInitiatedEventAttributes (Maybe Text)

The maximum duration for the child workflow execution. If the workflow execution isn't closed within this duration, it is timed out and force-terminated.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

startChildWorkflowExecutionInitiatedEventAttributes_taskStartToCloseTimeout :: Lens' StartChildWorkflowExecutionInitiatedEventAttributes (Maybe Text)

The maximum duration allowed for the decision tasks for this workflow execution.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

startChildWorkflowExecutionInitiatedEventAttributes_tagList :: Lens' StartChildWorkflowExecutionInitiatedEventAttributes (Maybe [Text])

The list of tags to associated with the child workflow execution.

startChildWorkflowExecutionInitiatedEventAttributes_workflowId :: Lens' StartChildWorkflowExecutionInitiatedEventAttributes Text

The workflowId of the child workflow execution.

startChildWorkflowExecutionInitiatedEventAttributes_workflowType :: Lens' StartChildWorkflowExecutionInitiatedEventAttributes WorkflowType

The type of the child workflow execution.

startChildWorkflowExecutionInitiatedEventAttributes_taskList :: Lens' StartChildWorkflowExecutionInitiatedEventAttributes TaskList

The name of the task list used for the decision tasks of the child workflow execution.

startChildWorkflowExecutionInitiatedEventAttributes_decisionTaskCompletedEventId :: Lens' StartChildWorkflowExecutionInitiatedEventAttributes Integer

The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the StartChildWorkflowExecution Decision to request this child workflow execution. This information can be useful for diagnosing problems by tracing back the cause of events.

startChildWorkflowExecutionInitiatedEventAttributes_childPolicy :: Lens' StartChildWorkflowExecutionInitiatedEventAttributes ChildPolicy

The policy to use for the child workflow executions if this execution gets terminated by explicitly calling the TerminateWorkflowExecution action or due to an expired timeout.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.