Network.AWS.SWF.Types.WorkflowExecution

Description

Documentation

data WorkflowExecution

Represents a workflow execution.

See: newWorkflowExecution smart constructor.

Constructors

WorkflowExecution'

Fields

workflowId :: Text

The user defined identifier associated with the workflow execution.

runId :: Text

A system-generated unique identifier for the workflow execution.

newWorkflowExecution

Arguments

:: Text
$sel:workflowId:WorkflowExecution'

-> Text
$sel:runId:WorkflowExecution'

-> WorkflowExecution

Create a value of WorkflowExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workflowId:WorkflowExecution', workflowExecution_workflowId - The user defined identifier associated with the workflow execution.

$sel:runId:WorkflowExecution', workflowExecution_runId - A system-generated unique identifier for the workflow execution.

workflowExecution_workflowId :: Lens' WorkflowExecution Text

The user defined identifier associated with the workflow execution.

workflowExecution_runId :: Lens' WorkflowExecution Text

A system-generated unique identifier for the workflow execution.