Network.AWS.SWF.Types.WorkflowExecutionContinuedAsNewEventAttributes

Description

Documentation

data WorkflowExecutionContinuedAsNewEventAttributes

Provides the details of the WorkflowExecutionContinuedAsNew event.

See: newWorkflowExecutionContinuedAsNewEventAttributes smart constructor.

Constructors

WorkflowExecutionContinuedAsNewEventAttributes'

Fields

tagList :: Maybe [Text]

The list of tags associated with the new workflow execution.

taskStartToCloseTimeout :: Maybe Text

The maximum duration of decision tasks for the new workflow execution.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

lambdaRole :: Maybe Text

The IAM role to attach to the new (continued) workflow execution.

input :: Maybe Text

The input provided to the new workflow execution.

executionStartToCloseTimeout :: Maybe Text

The total duration allowed for the new workflow execution.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

taskPriority :: Maybe Text

The priority of the task to use for the decisions of the new (continued) workflow execution.

decisionTaskCompletedEventId :: Integer

The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the ContinueAsNewWorkflowExecution decision that started this execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

newExecutionRunId' :: Text

The runId of the new workflow execution.

taskList :: TaskList

The task list to use for the decisions of the new (continued) workflow execution.

childPolicy :: ChildPolicy

The policy to use for the child workflow executions of the new execution if it is terminated by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

workflowType :: WorkflowType

The workflow type of this execution.

newWorkflowExecutionContinuedAsNewEventAttributes

Arguments

:: Integer
$sel:decisionTaskCompletedEventId:WorkflowExecutionContinuedAsNewEventAttributes'

-> Text
$sel:newExecutionRunId':WorkflowExecutionContinuedAsNewEventAttributes'

-> TaskList
$sel:taskList:WorkflowExecutionContinuedAsNewEventAttributes'

-> ChildPolicy
$sel:childPolicy:WorkflowExecutionContinuedAsNewEventAttributes'

-> WorkflowType
$sel:workflowType:WorkflowExecutionContinuedAsNewEventAttributes'

-> WorkflowExecutionContinuedAsNewEventAttributes

Create a value of WorkflowExecutionContinuedAsNewEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagList:WorkflowExecutionContinuedAsNewEventAttributes', workflowExecutionContinuedAsNewEventAttributes_tagList - The list of tags associated with the new workflow execution.

$sel:taskStartToCloseTimeout:WorkflowExecutionContinuedAsNewEventAttributes', workflowExecutionContinuedAsNewEventAttributes_taskStartToCloseTimeout - The maximum duration of decision tasks for the new workflow execution.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

$sel:lambdaRole:WorkflowExecutionContinuedAsNewEventAttributes', workflowExecutionContinuedAsNewEventAttributes_lambdaRole - The IAM role to attach to the new (continued) workflow execution.

$sel:input:WorkflowExecutionContinuedAsNewEventAttributes', workflowExecutionContinuedAsNewEventAttributes_input - The input provided to the new workflow execution.

$sel:executionStartToCloseTimeout:WorkflowExecutionContinuedAsNewEventAttributes', workflowExecutionContinuedAsNewEventAttributes_executionStartToCloseTimeout - The total duration allowed for the new workflow execution.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

$sel:taskPriority:WorkflowExecutionContinuedAsNewEventAttributes', workflowExecutionContinuedAsNewEventAttributes_taskPriority - The priority of the task to use for the decisions of the new (continued) workflow execution.

$sel:decisionTaskCompletedEventId:WorkflowExecutionContinuedAsNewEventAttributes', workflowExecutionContinuedAsNewEventAttributes_decisionTaskCompletedEventId - The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the ContinueAsNewWorkflowExecution decision that started this execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

$sel:newExecutionRunId':WorkflowExecutionContinuedAsNewEventAttributes', workflowExecutionContinuedAsNewEventAttributes_newExecutionRunId - The runId of the new workflow execution.

$sel:taskList:WorkflowExecutionContinuedAsNewEventAttributes', workflowExecutionContinuedAsNewEventAttributes_taskList - The task list to use for the decisions of the new (continued) workflow execution.

$sel:childPolicy:WorkflowExecutionContinuedAsNewEventAttributes', workflowExecutionContinuedAsNewEventAttributes_childPolicy - The policy to use for the child workflow executions of the new execution if it is terminated by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

$sel:workflowType:WorkflowExecutionContinuedAsNewEventAttributes', workflowExecutionContinuedAsNewEventAttributes_workflowType - The workflow type of this execution.

workflowExecutionContinuedAsNewEventAttributes_tagList :: Lens' WorkflowExecutionContinuedAsNewEventAttributes (Maybe [Text])

The list of tags associated with the new workflow execution.

workflowExecutionContinuedAsNewEventAttributes_taskStartToCloseTimeout :: Lens' WorkflowExecutionContinuedAsNewEventAttributes (Maybe Text)

The maximum duration of decision tasks for the new workflow execution.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

workflowExecutionContinuedAsNewEventAttributes_lambdaRole :: Lens' WorkflowExecutionContinuedAsNewEventAttributes (Maybe Text)

The IAM role to attach to the new (continued) workflow execution.

workflowExecutionContinuedAsNewEventAttributes_input :: Lens' WorkflowExecutionContinuedAsNewEventAttributes (Maybe Text)

The input provided to the new workflow execution.

workflowExecutionContinuedAsNewEventAttributes_executionStartToCloseTimeout :: Lens' WorkflowExecutionContinuedAsNewEventAttributes (Maybe Text)

The total duration allowed for the new workflow execution.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

workflowExecutionContinuedAsNewEventAttributes_taskPriority :: Lens' WorkflowExecutionContinuedAsNewEventAttributes (Maybe Text)

The priority of the task to use for the decisions of the new (continued) workflow execution.

workflowExecutionContinuedAsNewEventAttributes_decisionTaskCompletedEventId :: Lens' WorkflowExecutionContinuedAsNewEventAttributes Integer

The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the ContinueAsNewWorkflowExecution decision that started this execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

workflowExecutionContinuedAsNewEventAttributes_newExecutionRunId :: Lens' WorkflowExecutionContinuedAsNewEventAttributes Text

The runId of the new workflow execution.

workflowExecutionContinuedAsNewEventAttributes_taskList :: Lens' WorkflowExecutionContinuedAsNewEventAttributes TaskList

The task list to use for the decisions of the new (continued) workflow execution.

workflowExecutionContinuedAsNewEventAttributes_childPolicy :: Lens' WorkflowExecutionContinuedAsNewEventAttributes ChildPolicy

The policy to use for the child workflow executions of the new execution if it is terminated by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

workflowExecutionContinuedAsNewEventAttributes_workflowType :: Lens' WorkflowExecutionContinuedAsNewEventAttributes WorkflowType

The workflow type of this execution.