Network.AWS.SWF.Types.WorkflowExecutionCount

Description

Documentation

data WorkflowExecutionCount

Contains the count of workflow executions returned from CountOpenWorkflowExecutions or CountClosedWorkflowExecutions

See: newWorkflowExecutionCount smart constructor.

Constructors

WorkflowExecutionCount'

Fields

truncated :: Maybe Bool

If set to true, indicates that the actual count was more than the maximum supported by this API and the count returned is the truncated value.

count :: Natural

The number of workflow executions.

newWorkflowExecutionCount

Arguments

:: Natural
$sel:count:WorkflowExecutionCount'

-> WorkflowExecutionCount

Create a value of WorkflowExecutionCount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:truncated:WorkflowExecutionCount', workflowExecutionCount_truncated - If set to true, indicates that the actual count was more than the maximum supported by this API and the count returned is the truncated value.

$sel:count:WorkflowExecutionCount', workflowExecutionCount_count - The number of workflow executions.

workflowExecutionCount_truncated :: Lens' WorkflowExecutionCount (Maybe Bool)

If set to true, indicates that the actual count was more than the maximum supported by this API and the count returned is the truncated value.

workflowExecutionCount_count :: Lens' WorkflowExecutionCount Natural

The number of workflow executions.