Network.AWS.SWF.Types.WorkflowExecutionInfo

Description

Documentation

data WorkflowExecutionInfo

Contains information about a workflow execution.

See: newWorkflowExecutionInfo smart constructor.

Constructors

WorkflowExecutionInfo'

Fields

parent :: Maybe WorkflowExecution

If this workflow execution is a child of another execution then contains the workflow execution that started this execution.

closeStatus :: Maybe CloseStatus

If the execution status is closed then this specifies how the execution was closed:

COMPLETED – the execution was successfully completed.

CANCELED – the execution was canceled.Cancellation allows the implementation to gracefully clean up before the execution is closed.

TERMINATED – the execution was force terminated.

FAILED – the execution failed to complete.

TIMED_OUT – the execution did not complete in the alloted time and was automatically timed out.

CONTINUED_AS_NEW – the execution is logically continued. This means the current execution was completed and a new execution was started to carry on the workflow.

cancelRequested :: Maybe Bool

Set to true if a cancellation is requested for this workflow execution.

closeTimestamp :: Maybe POSIX

The time when the workflow execution was closed. Set only if the execution status is CLOSED.

tagList :: Maybe [Text]

The list of tags associated with the workflow execution. Tags can be used to identify and list workflow executions of interest through the visibility APIs. A workflow execution can have a maximum of 5 tags.

execution :: WorkflowExecution

The workflow execution this information is about.

workflowType :: WorkflowType

The type of the workflow execution.

startTimestamp :: POSIX

The time when the execution was started.

executionStatus :: ExecutionStatus

The current status of the execution.

newWorkflowExecutionInfo

Arguments

:: WorkflowExecution
$sel:execution:WorkflowExecutionInfo'

-> WorkflowType
$sel:workflowType:WorkflowExecutionInfo'

-> UTCTime
$sel:startTimestamp:WorkflowExecutionInfo'

-> ExecutionStatus
$sel:executionStatus:WorkflowExecutionInfo'

-> WorkflowExecutionInfo

Create a value of WorkflowExecutionInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parent:WorkflowExecutionInfo', workflowExecutionInfo_parent - If this workflow execution is a child of another execution then contains the workflow execution that started this execution.

$sel:closeStatus:WorkflowExecutionInfo', workflowExecutionInfo_closeStatus - If the execution status is closed then this specifies how the execution was closed:

COMPLETED – the execution was successfully completed.

CANCELED – the execution was canceled.Cancellation allows the implementation to gracefully clean up before the execution is closed.

TERMINATED – the execution was force terminated.

FAILED – the execution failed to complete.

TIMED_OUT – the execution did not complete in the alloted time and was automatically timed out.

CONTINUED_AS_NEW – the execution is logically continued. This means the current execution was completed and a new execution was started to carry on the workflow.

$sel:cancelRequested:WorkflowExecutionInfo', workflowExecutionInfo_cancelRequested - Set to true if a cancellation is requested for this workflow execution.

$sel:closeTimestamp:WorkflowExecutionInfo', workflowExecutionInfo_closeTimestamp - The time when the workflow execution was closed. Set only if the execution status is CLOSED.

$sel:tagList:WorkflowExecutionInfo', workflowExecutionInfo_tagList - The list of tags associated with the workflow execution. Tags can be used to identify and list workflow executions of interest through the visibility APIs. A workflow execution can have a maximum of 5 tags.

$sel:execution:WorkflowExecutionInfo', workflowExecutionInfo_execution - The workflow execution this information is about.

$sel:workflowType:WorkflowExecutionInfo', workflowExecutionInfo_workflowType - The type of the workflow execution.

$sel:startTimestamp:WorkflowExecutionInfo', workflowExecutionInfo_startTimestamp - The time when the execution was started.

$sel:executionStatus:WorkflowExecutionInfo', workflowExecutionInfo_executionStatus - The current status of the execution.

workflowExecutionInfo_parent :: Lens' WorkflowExecutionInfo (Maybe WorkflowExecution)

If this workflow execution is a child of another execution then contains the workflow execution that started this execution.

workflowExecutionInfo_closeStatus :: Lens' WorkflowExecutionInfo (Maybe CloseStatus)

If the execution status is closed then this specifies how the execution was closed:

COMPLETED – the execution was successfully completed.

CANCELED – the execution was canceled.Cancellation allows the implementation to gracefully clean up before the execution is closed.

TERMINATED – the execution was force terminated.

FAILED – the execution failed to complete.

TIMED_OUT – the execution did not complete in the alloted time and was automatically timed out.

CONTINUED_AS_NEW – the execution is logically continued. This means the current execution was completed and a new execution was started to carry on the workflow.

workflowExecutionInfo_cancelRequested :: Lens' WorkflowExecutionInfo (Maybe Bool)

Set to true if a cancellation is requested for this workflow execution.

workflowExecutionInfo_closeTimestamp :: Lens' WorkflowExecutionInfo (Maybe UTCTime)

The time when the workflow execution was closed. Set only if the execution status is CLOSED.

workflowExecutionInfo_tagList :: Lens' WorkflowExecutionInfo (Maybe [Text])

The list of tags associated with the workflow execution. Tags can be used to identify and list workflow executions of interest through the visibility APIs. A workflow execution can have a maximum of 5 tags.

workflowExecutionInfo_execution :: Lens' WorkflowExecutionInfo WorkflowExecution

The workflow execution this information is about.

workflowExecutionInfo_workflowType :: Lens' WorkflowExecutionInfo WorkflowType

The type of the workflow execution.

workflowExecutionInfo_startTimestamp :: Lens' WorkflowExecutionInfo UTCTime

The time when the execution was started.

workflowExecutionInfo_executionStatus :: Lens' WorkflowExecutionInfo ExecutionStatus

The current status of the execution.