Network.AWS.SWF.Types.WorkflowExecutionOpenCounts

Description

Documentation

data WorkflowExecutionOpenCounts

Contains the counts of open tasks, child workflow executions and timers for a workflow execution.

See: newWorkflowExecutionOpenCounts smart constructor.

Constructors

WorkflowExecutionOpenCounts'

Fields

openLambdaFunctions :: Maybe Natural

The count of Lambda tasks whose status is OPEN.

openActivityTasks :: Natural

The count of activity tasks whose status is OPEN.

openDecisionTasks :: Natural

The count of decision tasks whose status is OPEN. A workflow execution can have at most one open decision task.

openTimers :: Natural

The count of timers started by this workflow execution that have not fired yet.

openChildWorkflowExecutions :: Natural

The count of child workflow executions whose status is OPEN.

newWorkflowExecutionOpenCounts

Arguments

:: Natural
$sel:openActivityTasks:WorkflowExecutionOpenCounts'

-> Natural
$sel:openDecisionTasks:WorkflowExecutionOpenCounts'

-> Natural
$sel:openTimers:WorkflowExecutionOpenCounts'

-> Natural
$sel:openChildWorkflowExecutions:WorkflowExecutionOpenCounts'

-> WorkflowExecutionOpenCounts

Create a value of WorkflowExecutionOpenCounts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:openLambdaFunctions:WorkflowExecutionOpenCounts', workflowExecutionOpenCounts_openLambdaFunctions - The count of Lambda tasks whose status is OPEN.

$sel:openActivityTasks:WorkflowExecutionOpenCounts', workflowExecutionOpenCounts_openActivityTasks - The count of activity tasks whose status is OPEN.

$sel:openDecisionTasks:WorkflowExecutionOpenCounts', workflowExecutionOpenCounts_openDecisionTasks - The count of decision tasks whose status is OPEN. A workflow execution can have at most one open decision task.

$sel:openTimers:WorkflowExecutionOpenCounts', workflowExecutionOpenCounts_openTimers - The count of timers started by this workflow execution that have not fired yet.

$sel:openChildWorkflowExecutions:WorkflowExecutionOpenCounts', workflowExecutionOpenCounts_openChildWorkflowExecutions - The count of child workflow executions whose status is OPEN.

workflowExecutionOpenCounts_openLambdaFunctions :: Lens' WorkflowExecutionOpenCounts (Maybe Natural)

The count of Lambda tasks whose status is OPEN.

workflowExecutionOpenCounts_openActivityTasks :: Lens' WorkflowExecutionOpenCounts Natural

The count of activity tasks whose status is OPEN.

workflowExecutionOpenCounts_openDecisionTasks :: Lens' WorkflowExecutionOpenCounts Natural

The count of decision tasks whose status is OPEN. A workflow execution can have at most one open decision task.

workflowExecutionOpenCounts_openTimers :: Lens' WorkflowExecutionOpenCounts Natural

The count of timers started by this workflow execution that have not fired yet.

workflowExecutionOpenCounts_openChildWorkflowExecutions :: Lens' WorkflowExecutionOpenCounts Natural

The count of child workflow executions whose status is OPEN.