Network.AWS.SWF.Types.WorkflowExecutionStartedEventAttributes

Description

Documentation

data WorkflowExecutionStartedEventAttributes

Provides details of WorkflowExecutionStarted event.

See: newWorkflowExecutionStartedEventAttributes smart constructor.

Constructors

WorkflowExecutionStartedEventAttributes'

Fields

parentInitiatedEventId :: Maybe Integer

The ID of the StartChildWorkflowExecutionInitiated event corresponding to the StartChildWorkflowExecution Decision to start this workflow execution. The source event with this ID can be found in the history of the source workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

tagList :: Maybe [Text]

The list of tags associated with this workflow execution. An execution can have up to 5 tags.

taskStartToCloseTimeout :: Maybe Text

The maximum duration of decision tasks for this workflow type.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

lambdaRole :: Maybe Text

The IAM role attached to the workflow execution.

input :: Maybe Text

The input provided to the workflow execution.

executionStartToCloseTimeout :: Maybe Text

The maximum duration for this workflow execution.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

taskPriority :: Maybe Text

The priority of the decision tasks in the workflow execution.

parentWorkflowExecution :: Maybe WorkflowExecution

The source workflow execution that started this workflow execution. The member isn't set if the workflow execution was not started by a workflow.

continuedExecutionRunId :: Maybe Text

If this workflow execution was started due to a ContinueAsNewWorkflowExecution decision, then it contains the runId of the previous workflow execution that was closed and continued as this execution.

childPolicy :: ChildPolicy

The policy to use for the child workflow executions if this workflow execution is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

taskList :: TaskList

The name of the task list for scheduling the decision tasks for this workflow execution.

workflowType :: WorkflowType

The workflow type of this execution.

newWorkflowExecutionStartedEventAttributes

Arguments

:: ChildPolicy
$sel:childPolicy:WorkflowExecutionStartedEventAttributes'

-> TaskList
$sel:taskList:WorkflowExecutionStartedEventAttributes'

-> WorkflowType
$sel:workflowType:WorkflowExecutionStartedEventAttributes'

-> WorkflowExecutionStartedEventAttributes

Create a value of WorkflowExecutionStartedEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parentInitiatedEventId:WorkflowExecutionStartedEventAttributes', workflowExecutionStartedEventAttributes_parentInitiatedEventId - The ID of the StartChildWorkflowExecutionInitiated event corresponding to the StartChildWorkflowExecution Decision to start this workflow execution. The source event with this ID can be found in the history of the source workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

$sel:tagList:WorkflowExecutionStartedEventAttributes', workflowExecutionStartedEventAttributes_tagList - The list of tags associated with this workflow execution. An execution can have up to 5 tags.

$sel:taskStartToCloseTimeout:WorkflowExecutionStartedEventAttributes', workflowExecutionStartedEventAttributes_taskStartToCloseTimeout - The maximum duration of decision tasks for this workflow type.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

$sel:lambdaRole:WorkflowExecutionStartedEventAttributes', workflowExecutionStartedEventAttributes_lambdaRole - The IAM role attached to the workflow execution.

$sel:input:WorkflowExecutionStartedEventAttributes', workflowExecutionStartedEventAttributes_input - The input provided to the workflow execution.

$sel:executionStartToCloseTimeout:WorkflowExecutionStartedEventAttributes', workflowExecutionStartedEventAttributes_executionStartToCloseTimeout - The maximum duration for this workflow execution.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

$sel:taskPriority:WorkflowExecutionStartedEventAttributes', workflowExecutionStartedEventAttributes_taskPriority - The priority of the decision tasks in the workflow execution.

$sel:parentWorkflowExecution:WorkflowExecutionStartedEventAttributes', workflowExecutionStartedEventAttributes_parentWorkflowExecution - The source workflow execution that started this workflow execution. The member isn't set if the workflow execution was not started by a workflow.

$sel:continuedExecutionRunId:WorkflowExecutionStartedEventAttributes', workflowExecutionStartedEventAttributes_continuedExecutionRunId - If this workflow execution was started due to a ContinueAsNewWorkflowExecution decision, then it contains the runId of the previous workflow execution that was closed and continued as this execution.

$sel:childPolicy:WorkflowExecutionStartedEventAttributes', workflowExecutionStartedEventAttributes_childPolicy - The policy to use for the child workflow executions if this workflow execution is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

$sel:taskList:WorkflowExecutionStartedEventAttributes', workflowExecutionStartedEventAttributes_taskList - The name of the task list for scheduling the decision tasks for this workflow execution.

$sel:workflowType:WorkflowExecutionStartedEventAttributes', workflowExecutionStartedEventAttributes_workflowType - The workflow type of this execution.

workflowExecutionStartedEventAttributes_parentInitiatedEventId :: Lens' WorkflowExecutionStartedEventAttributes (Maybe Integer)

The ID of the StartChildWorkflowExecutionInitiated event corresponding to the StartChildWorkflowExecution Decision to start this workflow execution. The source event with this ID can be found in the history of the source workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

workflowExecutionStartedEventAttributes_tagList :: Lens' WorkflowExecutionStartedEventAttributes (Maybe [Text])

The list of tags associated with this workflow execution. An execution can have up to 5 tags.

workflowExecutionStartedEventAttributes_taskStartToCloseTimeout :: Lens' WorkflowExecutionStartedEventAttributes (Maybe Text)

The maximum duration of decision tasks for this workflow type.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

workflowExecutionStartedEventAttributes_lambdaRole :: Lens' WorkflowExecutionStartedEventAttributes (Maybe Text)

The IAM role attached to the workflow execution.

workflowExecutionStartedEventAttributes_input :: Lens' WorkflowExecutionStartedEventAttributes (Maybe Text)

The input provided to the workflow execution.

workflowExecutionStartedEventAttributes_executionStartToCloseTimeout :: Lens' WorkflowExecutionStartedEventAttributes (Maybe Text)

The maximum duration for this workflow execution.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

workflowExecutionStartedEventAttributes_taskPriority :: Lens' WorkflowExecutionStartedEventAttributes (Maybe Text)

The priority of the decision tasks in the workflow execution.

workflowExecutionStartedEventAttributes_parentWorkflowExecution :: Lens' WorkflowExecutionStartedEventAttributes (Maybe WorkflowExecution)

The source workflow execution that started this workflow execution. The member isn't set if the workflow execution was not started by a workflow.

workflowExecutionStartedEventAttributes_continuedExecutionRunId :: Lens' WorkflowExecutionStartedEventAttributes (Maybe Text)

If this workflow execution was started due to a ContinueAsNewWorkflowExecution decision, then it contains the runId of the previous workflow execution that was closed and continued as this execution.

workflowExecutionStartedEventAttributes_childPolicy :: Lens' WorkflowExecutionStartedEventAttributes ChildPolicy

The policy to use for the child workflow executions if this workflow execution is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

workflowExecutionStartedEventAttributes_taskList :: Lens' WorkflowExecutionStartedEventAttributes TaskList

The name of the task list for scheduling the decision tasks for this workflow execution.

workflowExecutionStartedEventAttributes_workflowType :: Lens' WorkflowExecutionStartedEventAttributes WorkflowType

The workflow type of this execution.