Network.AWS.SWF.Types.WorkflowTypeConfiguration

Description

Documentation

data WorkflowTypeConfiguration

The configuration settings of a workflow type.

See: newWorkflowTypeConfiguration smart constructor.

Constructors

WorkflowTypeConfiguration'

Fields

defaultExecutionStartToCloseTimeout :: Maybe Text

The default maximum duration, specified when registering the workflow type, for executions of this workflow type. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution Decision.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

defaultTaskPriority :: Maybe Text

The default task priority, specified when registering the workflow type, for all decision tasks of this workflow type. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution decision.

Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

defaultTaskList :: Maybe TaskList

The default task list, specified when registering the workflow type, for decisions tasks scheduled for workflow executions of this type. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution Decision.

defaultChildPolicy :: Maybe ChildPolicy

The default policy to use for the child workflow executions when a workflow execution of this type is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution Decision.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

defaultTaskStartToCloseTimeout :: Maybe Text

The default maximum duration, specified when registering the workflow type, that a decision task for executions of this workflow type might take before returning completion or failure. If the task doesn'tdo close in the specified time then the task is automatically timed out and rescheduled. If the decider eventually reports a completion or failure, it is ignored. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution Decision.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

defaultLambdaRole :: Maybe Text

The default IAM role attached to this workflow type.

Executions of this workflow type need IAM roles to invoke Lambda functions. If you don't specify an IAM role when starting this workflow type, the default Lambda role is attached to the execution. For more information, see https://docs.aws.amazon.com/amazonswf/latest/developerguide/lambda-task.html in the Amazon SWF Developer Guide.

newWorkflowTypeConfiguration :: WorkflowTypeConfiguration

Create a value of WorkflowTypeConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultExecutionStartToCloseTimeout:WorkflowTypeConfiguration', workflowTypeConfiguration_defaultExecutionStartToCloseTimeout - The default maximum duration, specified when registering the workflow type, for executions of this workflow type. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution Decision.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

$sel:defaultTaskPriority:WorkflowTypeConfiguration', workflowTypeConfiguration_defaultTaskPriority - The default task priority, specified when registering the workflow type, for all decision tasks of this workflow type. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution decision.

Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

$sel:defaultTaskList:WorkflowTypeConfiguration', workflowTypeConfiguration_defaultTaskList - The default task list, specified when registering the workflow type, for decisions tasks scheduled for workflow executions of this type. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution Decision.

$sel:defaultChildPolicy:WorkflowTypeConfiguration', workflowTypeConfiguration_defaultChildPolicy - The default policy to use for the child workflow executions when a workflow execution of this type is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution Decision.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

$sel:defaultTaskStartToCloseTimeout:WorkflowTypeConfiguration', workflowTypeConfiguration_defaultTaskStartToCloseTimeout - The default maximum duration, specified when registering the workflow type, that a decision task for executions of this workflow type might take before returning completion or failure. If the task doesn'tdo close in the specified time then the task is automatically timed out and rescheduled. If the decider eventually reports a completion or failure, it is ignored. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution Decision.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

$sel:defaultLambdaRole:WorkflowTypeConfiguration', workflowTypeConfiguration_defaultLambdaRole - The default IAM role attached to this workflow type.

Executions of this workflow type need IAM roles to invoke Lambda functions. If you don't specify an IAM role when starting this workflow type, the default Lambda role is attached to the execution. For more information, see https://docs.aws.amazon.com/amazonswf/latest/developerguide/lambda-task.html in the Amazon SWF Developer Guide.

workflowTypeConfiguration_defaultExecutionStartToCloseTimeout :: Lens' WorkflowTypeConfiguration (Maybe Text)

The default maximum duration, specified when registering the workflow type, for executions of this workflow type. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution Decision.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

workflowTypeConfiguration_defaultTaskPriority :: Lens' WorkflowTypeConfiguration (Maybe Text)

The default task priority, specified when registering the workflow type, for all decision tasks of this workflow type. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution decision.

Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

workflowTypeConfiguration_defaultTaskList :: Lens' WorkflowTypeConfiguration (Maybe TaskList)

The default task list, specified when registering the workflow type, for decisions tasks scheduled for workflow executions of this type. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution Decision.

workflowTypeConfiguration_defaultChildPolicy :: Lens' WorkflowTypeConfiguration (Maybe ChildPolicy)

The default policy to use for the child workflow executions when a workflow execution of this type is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution Decision.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

workflowTypeConfiguration_defaultTaskStartToCloseTimeout :: Lens' WorkflowTypeConfiguration (Maybe Text)

The default maximum duration, specified when registering the workflow type, that a decision task for executions of this workflow type might take before returning completion or failure. If the task doesn'tdo close in the specified time then the task is automatically timed out and rescheduled. If the decider eventually reports a completion or failure, it is ignored. This default can be overridden when starting a workflow execution using the StartWorkflowExecution action or the StartChildWorkflowExecution Decision.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

workflowTypeConfiguration_defaultLambdaRole :: Lens' WorkflowTypeConfiguration (Maybe Text)

The default IAM role attached to this workflow type.

Executions of this workflow type need IAM roles to invoke Lambda functions. If you don't specify an IAM role when starting this workflow type, the default Lambda role is attached to the execution. For more information, see https://docs.aws.amazon.com/amazonswf/latest/developerguide/lambda-task.html in the Amazon SWF Developer Guide.