Network.AWS.SWF.Types.WorkflowTypeInfo

Description

Documentation

data WorkflowTypeInfo

Contains information about a workflow type.

See: newWorkflowTypeInfo smart constructor.

Constructors

WorkflowTypeInfo'

Fields

deprecationDate :: Maybe POSIX

If the type is in deprecated state, then it is set to the date when the type was deprecated.

description :: Maybe Text

The description of the type registered through RegisterWorkflowType.

workflowType :: WorkflowType

The workflow type this information is about.

status :: RegistrationStatus

The current status of the workflow type.

creationDate :: POSIX

The date when this type was registered.

newWorkflowTypeInfo

Arguments

:: WorkflowType
$sel:workflowType:WorkflowTypeInfo'

-> RegistrationStatus
$sel:status:WorkflowTypeInfo'

-> UTCTime
$sel:creationDate:WorkflowTypeInfo'

-> WorkflowTypeInfo

Create a value of WorkflowTypeInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deprecationDate:WorkflowTypeInfo', workflowTypeInfo_deprecationDate - If the type is in deprecated state, then it is set to the date when the type was deprecated.

$sel:description:WorkflowTypeInfo', workflowTypeInfo_description - The description of the type registered through RegisterWorkflowType.

$sel:workflowType:WorkflowTypeInfo', workflowTypeInfo_workflowType - The workflow type this information is about.

$sel:status:WorkflowTypeInfo', workflowTypeInfo_status - The current status of the workflow type.

$sel:creationDate:WorkflowTypeInfo', workflowTypeInfo_creationDate - The date when this type was registered.

workflowTypeInfo_deprecationDate :: Lens' WorkflowTypeInfo (Maybe UTCTime)

If the type is in deprecated state, then it is set to the date when the type was deprecated.

workflowTypeInfo_description :: Lens' WorkflowTypeInfo (Maybe Text)

The description of the type registered through RegisterWorkflowType.

workflowTypeInfo_workflowType :: Lens' WorkflowTypeInfo WorkflowType

The workflow type this information is about.

workflowTypeInfo_status :: Lens' WorkflowTypeInfo RegistrationStatus

The current status of the workflow type.

workflowTypeInfo_creationDate :: Lens' WorkflowTypeInfo UTCTime

The date when this type was registered.