Network.AWS.SWF.UndeprecateActivityType

Description

Undeprecates a previously deprecated activity type. After an activity type has been undeprecated, you can create new tasks of that activity type.

This operation is eventually consistent. The results are best effort and may not exactly reflect recent updates and changes.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

activityType.name: String constraint. The key is swf:activityType.name.

activityType.version: String constraint. The key is swf:activityType.version.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

Creating a Request

data UndeprecateActivityType

See: newUndeprecateActivityType smart constructor.

Constructors

UndeprecateActivityType'

Fields

domain :: Text

The name of the domain of the deprecated activity type.

activityType :: ActivityType

The activity type to undeprecate.

newUndeprecateActivityType

Arguments

:: Text
$sel:domain:UndeprecateActivityType'

-> ActivityType
$sel:activityType:UndeprecateActivityType'

-> UndeprecateActivityType

Create a value of UndeprecateActivityType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domain:UndeprecateActivityType', undeprecateActivityType_domain - The name of the domain of the deprecated activity type.

$sel:activityType:UndeprecateActivityType', undeprecateActivityType_activityType - The activity type to undeprecate.

Request Lenses

undeprecateActivityType_domain :: Lens' UndeprecateActivityType Text

The name of the domain of the deprecated activity type.

undeprecateActivityType_activityType :: Lens' UndeprecateActivityType ActivityType

The activity type to undeprecate.

Destructuring the Response

data UndeprecateActivityTypeResponse

See: newUndeprecateActivityTypeResponse smart constructor.

Constructors

UndeprecateActivityTypeResponse'

newUndeprecateActivityTypeResponse :: UndeprecateActivityTypeResponse

Create a value of UndeprecateActivityTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.