Network.AWS.SWF.UndeprecateWorkflowType

Description

Undeprecates a previously deprecated workflow type. After a workflow type has been undeprecated, you can create new executions of that type.

This operation is eventually consistent. The results are best effort and may not exactly reflect recent updates and changes.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

workflowType.name: String constraint. The key is swf:workflowType.name.

workflowType.version: String constraint. The key is swf:workflowType.version.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

Creating a Request

data UndeprecateWorkflowType

See: newUndeprecateWorkflowType smart constructor.

Constructors

UndeprecateWorkflowType'

Fields

domain :: Text

The name of the domain of the deprecated workflow type.

workflowType :: WorkflowType

The name of the domain of the deprecated workflow type.

newUndeprecateWorkflowType

Arguments

:: Text
$sel:domain:UndeprecateWorkflowType'

-> WorkflowType
$sel:workflowType:UndeprecateWorkflowType'

-> UndeprecateWorkflowType

Create a value of UndeprecateWorkflowType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domain:UndeprecateWorkflowType', undeprecateWorkflowType_domain - The name of the domain of the deprecated workflow type.

$sel:workflowType:UndeprecateWorkflowType', undeprecateWorkflowType_workflowType - The name of the domain of the deprecated workflow type.

Request Lenses

undeprecateWorkflowType_domain :: Lens' UndeprecateWorkflowType Text

The name of the domain of the deprecated workflow type.

undeprecateWorkflowType_workflowType :: Lens' UndeprecateWorkflowType WorkflowType

The name of the domain of the deprecated workflow type.

Destructuring the Response

data UndeprecateWorkflowTypeResponse

See: newUndeprecateWorkflowTypeResponse smart constructor.

Constructors

UndeprecateWorkflowTypeResponse'

newUndeprecateWorkflowTypeResponse :: UndeprecateWorkflowTypeResponse

Create a value of UndeprecateWorkflowTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.