Network.AWS.SWF.V2012_01_25

Description

Amazon Simple Workflow Service (Amazon SWF) makes it easy to build applications that coordinate work across distributed components. In Amazon SWF, a task represents a logical unit of work that is performed by a component of your application. Coordinating tasks across the application involves managing intertask dependencies, scheduling, and concurrency in accordance with the logical flow of the application. Amazon SWF gives you full control over implementing tasks and coordinating them without worrying about underlying complexities such as tracking their progress and maintaining their state.

Documentation

module Network.AWS.SWF.V2012_01_25.CountClosedWorkflowExecutions

module Network.AWS.SWF.V2012_01_25.CountOpenWorkflowExecutions

module Network.AWS.SWF.V2012_01_25.CountPendingActivityTasks

module Network.AWS.SWF.V2012_01_25.CountPendingDecisionTasks

module Network.AWS.SWF.V2012_01_25.DeprecateActivityType

module Network.AWS.SWF.V2012_01_25.DeprecateDomain

module Network.AWS.SWF.V2012_01_25.DeprecateWorkflowType

module Network.AWS.SWF.V2012_01_25.DescribeActivityType

module Network.AWS.SWF.V2012_01_25.DescribeDomain

module Network.AWS.SWF.V2012_01_25.DescribeWorkflowExecution

module Network.AWS.SWF.V2012_01_25.DescribeWorkflowType

module Network.AWS.SWF.V2012_01_25.GetWorkflowExecutionHistory

module Network.AWS.SWF.V2012_01_25.ListActivityTypes

module Network.AWS.SWF.V2012_01_25.ListClosedWorkflowExecutions

module Network.AWS.SWF.V2012_01_25.ListDomains

module Network.AWS.SWF.V2012_01_25.ListOpenWorkflowExecutions

module Network.AWS.SWF.V2012_01_25.ListWorkflowTypes

module Network.AWS.SWF.V2012_01_25.PollForActivityTask

module Network.AWS.SWF.V2012_01_25.PollForDecisionTask

module Network.AWS.SWF.V2012_01_25.RecordActivityTaskHeartbeat

module Network.AWS.SWF.V2012_01_25.RegisterActivityType

module Network.AWS.SWF.V2012_01_25.RegisterDomain

module Network.AWS.SWF.V2012_01_25.RegisterWorkflowType

module Network.AWS.SWF.V2012_01_25.RequestCancelWorkflowExecution

module Network.AWS.SWF.V2012_01_25.RespondActivityTaskCanceled

module Network.AWS.SWF.V2012_01_25.RespondActivityTaskCompleted

module Network.AWS.SWF.V2012_01_25.RespondActivityTaskFailed

module Network.AWS.SWF.V2012_01_25.RespondDecisionTaskCompleted

module Network.AWS.SWF.V2012_01_25.SignalWorkflowExecution

module Network.AWS.SWF.V2012_01_25.StartWorkflowExecution

module Network.AWS.SWF.V2012_01_25.TerminateWorkflowExecution

module Network.AWS.SWF.V2012_01_25.Types