Network.AWS.SageMaker.AddAssociation

Description

Creates an association between the source and the destination. A source can be associated with multiple destinations, and a destination can be associated with multiple sources. An association is a lineage tracking entity. For more information, see Amazon SageMaker ML Lineage Tracking.

Creating a Request

data AddAssociation

See: newAddAssociation smart constructor.

Constructors

AddAssociation'

Fields

associationType :: Maybe AssociationEdgeType

The type of association. The following are suggested uses for each type. Amazon SageMaker places no restrictions on their use.

ContributedTo - The source contributed to the destination or had a part in enabling the destination. For example, the training data contributed to the training job.

AssociatedWith - The source is connected to the destination. For example, an approval workflow is associated with a model deployment.

DerivedFrom - The destination is a modification of the source. For example, a digest output of a channel input for a processing job is derived from the original inputs.

Produced - The source generated the destination. For example, a training job produced a model artifact.

sourceArn :: Text

The ARN of the source.

destinationArn :: Text

The Amazon Resource Name (ARN) of the destination.

newAddAssociation

Arguments

:: Text
$sel:sourceArn:AddAssociation'

-> Text
$sel:destinationArn:AddAssociation'

-> AddAssociation

Create a value of AddAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associationType:AddAssociation', addAssociation_associationType - The type of association. The following are suggested uses for each type. Amazon SageMaker places no restrictions on their use.

ContributedTo - The source contributed to the destination or had a part in enabling the destination. For example, the training data contributed to the training job.

AssociatedWith - The source is connected to the destination. For example, an approval workflow is associated with a model deployment.

DerivedFrom - The destination is a modification of the source. For example, a digest output of a channel input for a processing job is derived from the original inputs.

Produced - The source generated the destination. For example, a training job produced a model artifact.

$sel:sourceArn:AddAssociation', addAssociation_sourceArn - The ARN of the source.

$sel:destinationArn:AddAssociation', addAssociation_destinationArn - The Amazon Resource Name (ARN) of the destination.

Request Lenses

addAssociation_associationType :: Lens' AddAssociation (Maybe AssociationEdgeType)

The type of association. The following are suggested uses for each type. Amazon SageMaker places no restrictions on their use.

ContributedTo - The source contributed to the destination or had a part in enabling the destination. For example, the training data contributed to the training job.

AssociatedWith - The source is connected to the destination. For example, an approval workflow is associated with a model deployment.

DerivedFrom - The destination is a modification of the source. For example, a digest output of a channel input for a processing job is derived from the original inputs.

Produced - The source generated the destination. For example, a training job produced a model artifact.

addAssociation_sourceArn :: Lens' AddAssociation Text

The ARN of the source.

addAssociation_destinationArn :: Lens' AddAssociation Text

The Amazon Resource Name (ARN) of the destination.

Destructuring the Response

data AddAssociationResponse

See: newAddAssociationResponse smart constructor.

Constructors

AddAssociationResponse'

Fields

destinationArn :: Maybe Text

The Amazon Resource Name (ARN) of the destination.

sourceArn :: Maybe Text

The ARN of the source.

httpStatus :: Int

The response's http status code.

newAddAssociationResponse

Arguments

:: Int
$sel:httpStatus:AddAssociationResponse'

-> AddAssociationResponse

Create a value of AddAssociationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationArn:AddAssociation', addAssociationResponse_destinationArn - The Amazon Resource Name (ARN) of the destination.

$sel:sourceArn:AddAssociation', addAssociationResponse_sourceArn - The ARN of the source.

$sel:httpStatus:AddAssociationResponse', addAssociationResponse_httpStatus - The response's http status code.

Response Lenses

addAssociationResponse_destinationArn :: Lens' AddAssociationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the destination.

addAssociationResponse_sourceArn :: Lens' AddAssociationResponse (Maybe Text)

The ARN of the source.

addAssociationResponse_httpStatus :: Lens' AddAssociationResponse Int

The response's http status code.