Network.AWS.SageMaker.CreateAlgorithm

Description

Create a machine learning algorithm that you can use in Amazon SageMaker and list in the Amazon Web Services Marketplace.

Creating a Request

data CreateAlgorithm

See: newCreateAlgorithm smart constructor.

Constructors

CreateAlgorithm'

Fields

algorithmDescription :: Maybe Text

A description of the algorithm.

validationSpecification :: Maybe AlgorithmValidationSpecification

Specifies configurations for one or more training jobs and that Amazon SageMaker runs to test the algorithm's training code and, optionally, one or more batch transform jobs that Amazon SageMaker runs to test the algorithm's inference code.

certifyForMarketplace :: Maybe Bool

Whether to certify the algorithm so that it can be listed in Amazon Web Services Marketplace.

tags :: Maybe [Tag]

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

inferenceSpecification :: Maybe InferenceSpecification

Specifies details about inference jobs that the algorithm runs, including the following:

The Amazon ECR paths of containers that contain the inference code and model artifacts.

The instance types that the algorithm supports for transform jobs and real-time endpoints used for inference.

The input and output content formats that the algorithm supports for inference.

algorithmName :: Text

The name of the algorithm.

trainingSpecification :: TrainingSpecification

Specifies details about training jobs run by this algorithm, including the following:

The Amazon ECR path of the container and the version digest of the algorithm.

The hyperparameters that the algorithm supports.

The instance types that the algorithm supports for training.

Whether the algorithm supports distributed training.

The metrics that the algorithm emits to Amazon CloudWatch.

Which metrics that the algorithm emits can be used as the objective metric for hyperparameter tuning jobs.

The input channels that the algorithm supports for training data. For example, an algorithm might support train, validation, and test channels.

newCreateAlgorithm

Arguments

:: Text
$sel:algorithmName:CreateAlgorithm'

-> TrainingSpecification
$sel:trainingSpecification:CreateAlgorithm'

-> CreateAlgorithm

Create a value of CreateAlgorithm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:algorithmDescription:CreateAlgorithm', createAlgorithm_algorithmDescription - A description of the algorithm.

$sel:validationSpecification:CreateAlgorithm', createAlgorithm_validationSpecification - Specifies configurations for one or more training jobs and that Amazon SageMaker runs to test the algorithm's training code and, optionally, one or more batch transform jobs that Amazon SageMaker runs to test the algorithm's inference code.

$sel:certifyForMarketplace:CreateAlgorithm', createAlgorithm_certifyForMarketplace - Whether to certify the algorithm so that it can be listed in Amazon Web Services Marketplace.

$sel:tags:CreateAlgorithm', createAlgorithm_tags - An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

$sel:inferenceSpecification:CreateAlgorithm', createAlgorithm_inferenceSpecification - Specifies details about inference jobs that the algorithm runs, including the following:

The Amazon ECR paths of containers that contain the inference code and model artifacts.

The instance types that the algorithm supports for transform jobs and real-time endpoints used for inference.

The input and output content formats that the algorithm supports for inference.

$sel:algorithmName:CreateAlgorithm', createAlgorithm_algorithmName - The name of the algorithm.

$sel:trainingSpecification:CreateAlgorithm', createAlgorithm_trainingSpecification - Specifies details about training jobs run by this algorithm, including the following:

The Amazon ECR path of the container and the version digest of the algorithm.

The hyperparameters that the algorithm supports.

The instance types that the algorithm supports for training.

Whether the algorithm supports distributed training.

The metrics that the algorithm emits to Amazon CloudWatch.

Which metrics that the algorithm emits can be used as the objective metric for hyperparameter tuning jobs.

The input channels that the algorithm supports for training data. For example, an algorithm might support train, validation, and test channels.

Request Lenses

createAlgorithm_algorithmDescription :: Lens' CreateAlgorithm (Maybe Text)

A description of the algorithm.

createAlgorithm_validationSpecification :: Lens' CreateAlgorithm (Maybe AlgorithmValidationSpecification)

Specifies configurations for one or more training jobs and that Amazon SageMaker runs to test the algorithm's training code and, optionally, one or more batch transform jobs that Amazon SageMaker runs to test the algorithm's inference code.

createAlgorithm_certifyForMarketplace :: Lens' CreateAlgorithm (Maybe Bool)

Whether to certify the algorithm so that it can be listed in Amazon Web Services Marketplace.

createAlgorithm_tags :: Lens' CreateAlgorithm (Maybe [Tag])

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

createAlgorithm_inferenceSpecification :: Lens' CreateAlgorithm (Maybe InferenceSpecification)

Specifies details about inference jobs that the algorithm runs, including the following:

The Amazon ECR paths of containers that contain the inference code and model artifacts.

The instance types that the algorithm supports for transform jobs and real-time endpoints used for inference.

The input and output content formats that the algorithm supports for inference.

createAlgorithm_algorithmName :: Lens' CreateAlgorithm Text

The name of the algorithm.

createAlgorithm_trainingSpecification :: Lens' CreateAlgorithm TrainingSpecification

Specifies details about training jobs run by this algorithm, including the following:

The Amazon ECR path of the container and the version digest of the algorithm.

The hyperparameters that the algorithm supports.

The instance types that the algorithm supports for training.

Whether the algorithm supports distributed training.

The metrics that the algorithm emits to Amazon CloudWatch.

Which metrics that the algorithm emits can be used as the objective metric for hyperparameter tuning jobs.

The input channels that the algorithm supports for training data. For example, an algorithm might support train, validation, and test channels.

Destructuring the Response

data CreateAlgorithmResponse

See: newCreateAlgorithmResponse smart constructor.

Constructors

CreateAlgorithmResponse'

Fields

httpStatus :: Int

The response's http status code.

algorithmArn :: Text

The Amazon Resource Name (ARN) of the new algorithm.

newCreateAlgorithmResponse

Arguments

:: Int
$sel:httpStatus:CreateAlgorithmResponse'

-> Text
$sel:algorithmArn:CreateAlgorithmResponse'

-> CreateAlgorithmResponse

Create a value of CreateAlgorithmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateAlgorithmResponse', createAlgorithmResponse_httpStatus - The response's http status code.

$sel:algorithmArn:CreateAlgorithmResponse', createAlgorithmResponse_algorithmArn - The Amazon Resource Name (ARN) of the new algorithm.

Response Lenses

createAlgorithmResponse_httpStatus :: Lens' CreateAlgorithmResponse Int

The response's http status code.

createAlgorithmResponse_algorithmArn :: Lens' CreateAlgorithmResponse Text

The Amazon Resource Name (ARN) of the new algorithm.