Network.AWS.SageMaker.CreateAutoMLJob

Description

Creates an Autopilot job.

Find the best-performing model after you run an Autopilot job by calling .

For information about how to use Autopilot, see Automate Model Development with Amazon SageMaker Autopilot.

Creating a Request

data CreateAutoMLJob

See: newCreateAutoMLJob smart constructor.

Constructors

CreateAutoMLJob'

Fields

generateCandidateDefinitionsOnly :: Maybe Bool

Generates possible candidates without training the models. A candidate is a combination of data preprocessors, algorithms, and algorithm parameter settings.

tags :: Maybe [Tag]

Each tag consists of a key and an optional value. Tag keys must be unique per resource.

autoMLJobObjective :: Maybe AutoMLJobObjective

Defines the objective metric used to measure the predictive quality of an AutoML job. You provide an AutoMLJobObjective$MetricName and Autopilot infers whether to minimize or maximize it.

modelDeployConfig :: Maybe ModelDeployConfig

Specifies how to generate the endpoint name for an automatic one-click Autopilot model deployment.

autoMLJobConfig :: Maybe AutoMLJobConfig

Contains CompletionCriteria and SecurityConfig settings for the AutoML job.

problemType :: Maybe ProblemType

Defines the type of supervised learning available for the candidates. Options include: BinaryClassification, MulticlassClassification, and Regression. For more information, see Amazon SageMaker Autopilot problem types and algorithm support.

autoMLJobName :: Text

Identifies an Autopilot job. The name must be unique to your account and is case-insensitive.

inputDataConfig :: NonEmpty AutoMLChannel

An array of channel objects that describes the input data and its location. Each channel is a named input source. Similar to InputDataConfig supported by . Format(s) supported: CSV. Minimum of 500 rows.

outputDataConfig :: AutoMLOutputDataConfig

Provides information about encryption and the Amazon S3 output path needed to store artifacts from an AutoML job. Format(s) supported: CSV.

roleArn :: Text

The ARN of the role that is used to access the data.

newCreateAutoMLJob

Arguments

:: Text
$sel:autoMLJobName:CreateAutoMLJob'

-> NonEmpty AutoMLChannel
$sel:inputDataConfig:CreateAutoMLJob'

-> AutoMLOutputDataConfig
$sel:outputDataConfig:CreateAutoMLJob'

-> Text
$sel:roleArn:CreateAutoMLJob'

-> CreateAutoMLJob

Create a value of CreateAutoMLJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:generateCandidateDefinitionsOnly:CreateAutoMLJob', createAutoMLJob_generateCandidateDefinitionsOnly - Generates possible candidates without training the models. A candidate is a combination of data preprocessors, algorithms, and algorithm parameter settings.

$sel:tags:CreateAutoMLJob', createAutoMLJob_tags - Each tag consists of a key and an optional value. Tag keys must be unique per resource.

$sel:autoMLJobObjective:CreateAutoMLJob', createAutoMLJob_autoMLJobObjective - Defines the objective metric used to measure the predictive quality of an AutoML job. You provide an AutoMLJobObjective$MetricName and Autopilot infers whether to minimize or maximize it.

$sel:modelDeployConfig:CreateAutoMLJob', createAutoMLJob_modelDeployConfig - Specifies how to generate the endpoint name for an automatic one-click Autopilot model deployment.

$sel:autoMLJobConfig:CreateAutoMLJob', createAutoMLJob_autoMLJobConfig - Contains CompletionCriteria and SecurityConfig settings for the AutoML job.

$sel:problemType:CreateAutoMLJob', createAutoMLJob_problemType - Defines the type of supervised learning available for the candidates. Options include: BinaryClassification, MulticlassClassification, and Regression. For more information, see Amazon SageMaker Autopilot problem types and algorithm support.

$sel:autoMLJobName:CreateAutoMLJob', createAutoMLJob_autoMLJobName - Identifies an Autopilot job. The name must be unique to your account and is case-insensitive.

$sel:inputDataConfig:CreateAutoMLJob', createAutoMLJob_inputDataConfig - An array of channel objects that describes the input data and its location. Each channel is a named input source. Similar to InputDataConfig supported by . Format(s) supported: CSV. Minimum of 500 rows.

$sel:outputDataConfig:CreateAutoMLJob', createAutoMLJob_outputDataConfig - Provides information about encryption and the Amazon S3 output path needed to store artifacts from an AutoML job. Format(s) supported: CSV.

$sel:roleArn:CreateAutoMLJob', createAutoMLJob_roleArn - The ARN of the role that is used to access the data.

Request Lenses

createAutoMLJob_generateCandidateDefinitionsOnly :: Lens' CreateAutoMLJob (Maybe Bool)

Generates possible candidates without training the models. A candidate is a combination of data preprocessors, algorithms, and algorithm parameter settings.

createAutoMLJob_tags :: Lens' CreateAutoMLJob (Maybe [Tag])

Each tag consists of a key and an optional value. Tag keys must be unique per resource.

createAutoMLJob_autoMLJobObjective :: Lens' CreateAutoMLJob (Maybe AutoMLJobObjective)

Defines the objective metric used to measure the predictive quality of an AutoML job. You provide an AutoMLJobObjective$MetricName and Autopilot infers whether to minimize or maximize it.

createAutoMLJob_modelDeployConfig :: Lens' CreateAutoMLJob (Maybe ModelDeployConfig)

Specifies how to generate the endpoint name for an automatic one-click Autopilot model deployment.

createAutoMLJob_autoMLJobConfig :: Lens' CreateAutoMLJob (Maybe AutoMLJobConfig)

Contains CompletionCriteria and SecurityConfig settings for the AutoML job.

createAutoMLJob_problemType :: Lens' CreateAutoMLJob (Maybe ProblemType)

Defines the type of supervised learning available for the candidates. Options include: BinaryClassification, MulticlassClassification, and Regression. For more information, see Amazon SageMaker Autopilot problem types and algorithm support.

createAutoMLJob_autoMLJobName :: Lens' CreateAutoMLJob Text

Identifies an Autopilot job. The name must be unique to your account and is case-insensitive.

createAutoMLJob_inputDataConfig :: Lens' CreateAutoMLJob (NonEmpty AutoMLChannel)

An array of channel objects that describes the input data and its location. Each channel is a named input source. Similar to InputDataConfig supported by . Format(s) supported: CSV. Minimum of 500 rows.

createAutoMLJob_outputDataConfig :: Lens' CreateAutoMLJob AutoMLOutputDataConfig

Provides information about encryption and the Amazon S3 output path needed to store artifacts from an AutoML job. Format(s) supported: CSV.

createAutoMLJob_roleArn :: Lens' CreateAutoMLJob Text

The ARN of the role that is used to access the data.

Destructuring the Response

data CreateAutoMLJobResponse

See: newCreateAutoMLJobResponse smart constructor.

Constructors

CreateAutoMLJobResponse'

Fields

httpStatus :: Int

The response's http status code.

autoMLJobArn :: Text

The unique ARN assigned to the AutoML job when it is created.

newCreateAutoMLJobResponse

Arguments

:: Int
$sel:httpStatus:CreateAutoMLJobResponse'

-> Text
$sel:autoMLJobArn:CreateAutoMLJobResponse'

-> CreateAutoMLJobResponse

Create a value of CreateAutoMLJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateAutoMLJobResponse', createAutoMLJobResponse_httpStatus - The response's http status code.

$sel:autoMLJobArn:CreateAutoMLJobResponse', createAutoMLJobResponse_autoMLJobArn - The unique ARN assigned to the AutoML job when it is created.

Response Lenses

createAutoMLJobResponse_httpStatus :: Lens' CreateAutoMLJobResponse Int

The response's http status code.

createAutoMLJobResponse_autoMLJobArn :: Lens' CreateAutoMLJobResponse Text

The unique ARN assigned to the AutoML job when it is created.