Network.AWS.SageMaker.CreateCodeRepository

Description

Creates a Git repository as a resource in your Amazon SageMaker account. You can associate the repository with notebook instances so that you can use Git source control for the notebooks you create. The Git repository is a resource in your Amazon SageMaker account, so it can be associated with more than one notebook instance, and it persists independently from the lifecycle of any notebook instances it is associated with.

The repository can be hosted either in Amazon Web Services CodeCommit or in any other Git repository.

Creating a Request

data CreateCodeRepository

See: newCreateCodeRepository smart constructor.

Constructors

CreateCodeRepository'

Fields

tags :: Maybe [Tag]

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

codeRepositoryName :: Text

The name of the Git repository. The name must have 1 to 63 characters. Valid characters are a-z, A-Z, 0-9, and - (hyphen).

gitConfig :: GitConfig

Specifies details about the repository, including the URL where the repository is located, the default branch, and credentials to use to access the repository.

newCreateCodeRepository

Arguments

:: Text
$sel:codeRepositoryName:CreateCodeRepository'

-> GitConfig
$sel:gitConfig:CreateCodeRepository'

-> CreateCodeRepository

Create a value of CreateCodeRepository with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateCodeRepository', createCodeRepository_tags - An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

$sel:codeRepositoryName:CreateCodeRepository', createCodeRepository_codeRepositoryName - The name of the Git repository. The name must have 1 to 63 characters. Valid characters are a-z, A-Z, 0-9, and - (hyphen).

$sel:gitConfig:CreateCodeRepository', createCodeRepository_gitConfig - Specifies details about the repository, including the URL where the repository is located, the default branch, and credentials to use to access the repository.

Request Lenses

createCodeRepository_tags :: Lens' CreateCodeRepository (Maybe [Tag])

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

createCodeRepository_codeRepositoryName :: Lens' CreateCodeRepository Text

The name of the Git repository. The name must have 1 to 63 characters. Valid characters are a-z, A-Z, 0-9, and - (hyphen).

createCodeRepository_gitConfig :: Lens' CreateCodeRepository GitConfig

Specifies details about the repository, including the URL where the repository is located, the default branch, and credentials to use to access the repository.

Destructuring the Response

data CreateCodeRepositoryResponse

See: newCreateCodeRepositoryResponse smart constructor.

Constructors

CreateCodeRepositoryResponse'

Fields

httpStatus :: Int

The response's http status code.

codeRepositoryArn :: Text

The Amazon Resource Name (ARN) of the new repository.

newCreateCodeRepositoryResponse

Arguments

:: Int
$sel:httpStatus:CreateCodeRepositoryResponse'

-> Text
$sel:codeRepositoryArn:CreateCodeRepositoryResponse'

-> CreateCodeRepositoryResponse

Create a value of CreateCodeRepositoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateCodeRepositoryResponse', createCodeRepositoryResponse_httpStatus - The response's http status code.

$sel:codeRepositoryArn:CreateCodeRepositoryResponse', createCodeRepositoryResponse_codeRepositoryArn - The Amazon Resource Name (ARN) of the new repository.

Response Lenses

createCodeRepositoryResponse_httpStatus :: Lens' CreateCodeRepositoryResponse Int

The response's http status code.

createCodeRepositoryResponse_codeRepositoryArn :: Lens' CreateCodeRepositoryResponse Text

The Amazon Resource Name (ARN) of the new repository.