Network.AWS.SageMaker.CreateContext

Description

Creates a context. A context is a lineage tracking entity that represents a logical grouping of other tracking or experiment entities. Some examples are an endpoint and a model package. For more information, see Amazon SageMaker ML Lineage Tracking.

Creating a Request

data CreateContext

See: newCreateContext smart constructor.

Constructors

CreateContext'

Fields

tags :: Maybe [Tag]

A list of tags to apply to the context.

properties :: Maybe (HashMap Text Text)

A list of properties to add to the context.

description :: Maybe Text

The description of the context.

contextName :: Text

The name of the context. Must be unique to your account in an Amazon Web Services Region.

source :: ContextSource

The source type, ID, and URI.

contextType :: Text

The context type.

newCreateContext

Arguments

:: Text
$sel:contextName:CreateContext'

-> ContextSource
$sel:source:CreateContext'

-> Text
$sel:contextType:CreateContext'

-> CreateContext

Create a value of CreateContext with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateContext', createContext_tags - A list of tags to apply to the context.

$sel:properties:CreateContext', createContext_properties - A list of properties to add to the context.

$sel:description:CreateContext', createContext_description - The description of the context.

$sel:contextName:CreateContext', createContext_contextName - The name of the context. Must be unique to your account in an Amazon Web Services Region.

$sel:source:CreateContext', createContext_source - The source type, ID, and URI.

$sel:contextType:CreateContext', createContext_contextType - The context type.

Request Lenses

createContext_tags :: Lens' CreateContext (Maybe [Tag])

A list of tags to apply to the context.

createContext_properties :: Lens' CreateContext (Maybe (HashMap Text Text))

A list of properties to add to the context.

createContext_description :: Lens' CreateContext (Maybe Text)

The description of the context.

createContext_contextName :: Lens' CreateContext Text

The name of the context. Must be unique to your account in an Amazon Web Services Region.

createContext_source :: Lens' CreateContext ContextSource

The source type, ID, and URI.

createContext_contextType :: Lens' CreateContext Text

The context type.

Destructuring the Response

data CreateContextResponse

See: newCreateContextResponse smart constructor.

Constructors

CreateContextResponse'

Fields

contextArn :: Maybe Text

The Amazon Resource Name (ARN) of the context.

httpStatus :: Int

The response's http status code.

newCreateContextResponse

Arguments

:: Int
$sel:httpStatus:CreateContextResponse'

-> CreateContextResponse

Create a value of CreateContextResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contextArn:CreateContextResponse', createContextResponse_contextArn - The Amazon Resource Name (ARN) of the context.

$sel:httpStatus:CreateContextResponse', createContextResponse_httpStatus - The response's http status code.

Response Lenses

createContextResponse_contextArn :: Lens' CreateContextResponse (Maybe Text)

The Amazon Resource Name (ARN) of the context.

createContextResponse_httpStatus :: Lens' CreateContextResponse Int

The response's http status code.