Network.AWS.SageMaker.CreateDataQualityJobDefinition

Description

Creates a definition for a job that monitors data quality and drift. For information about model monitor, see Amazon SageMaker Model Monitor.

Creating a Request

data CreateDataQualityJobDefinition

See: newCreateDataQualityJobDefinition smart constructor.

Constructors

CreateDataQualityJobDefinition'

Fields

networkConfig :: Maybe MonitoringNetworkConfig

Specifies networking configuration for the monitoring job.

dataQualityBaselineConfig :: Maybe DataQualityBaselineConfig

Configures the constraints and baselines for the monitoring job.

tags :: Maybe [Tag]

(Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

stoppingCondition :: Maybe MonitoringStoppingCondition

jobDefinitionName :: Text

The name for the monitoring job definition.

dataQualityAppSpecification :: DataQualityAppSpecification

Specifies the container that runs the monitoring job.

dataQualityJobInput :: DataQualityJobInput

A list of inputs for the monitoring job. Currently endpoints are supported as monitoring inputs.

dataQualityJobOutputConfig :: MonitoringOutputConfig

jobResources :: MonitoringResources

roleArn :: Text

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

newCreateDataQualityJobDefinition

Arguments

:: Text
$sel:jobDefinitionName:CreateDataQualityJobDefinition'

-> DataQualityAppSpecification
$sel:dataQualityAppSpecification:CreateDataQualityJobDefinition'

-> DataQualityJobInput
$sel:dataQualityJobInput:CreateDataQualityJobDefinition'

-> MonitoringOutputConfig
$sel:dataQualityJobOutputConfig:CreateDataQualityJobDefinition'

-> MonitoringResources
$sel:jobResources:CreateDataQualityJobDefinition'

-> Text
$sel:roleArn:CreateDataQualityJobDefinition'

-> CreateDataQualityJobDefinition

Create a value of CreateDataQualityJobDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkConfig:CreateDataQualityJobDefinition', createDataQualityJobDefinition_networkConfig - Specifies networking configuration for the monitoring job.

$sel:dataQualityBaselineConfig:CreateDataQualityJobDefinition', createDataQualityJobDefinition_dataQualityBaselineConfig - Configures the constraints and baselines for the monitoring job.

$sel:tags:CreateDataQualityJobDefinition', createDataQualityJobDefinition_tags - (Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

$sel:stoppingCondition:CreateDataQualityJobDefinition', createDataQualityJobDefinition_stoppingCondition - Undocumented member.

$sel:jobDefinitionName:CreateDataQualityJobDefinition', createDataQualityJobDefinition_jobDefinitionName - The name for the monitoring job definition.

$sel:dataQualityAppSpecification:CreateDataQualityJobDefinition', createDataQualityJobDefinition_dataQualityAppSpecification - Specifies the container that runs the monitoring job.

$sel:dataQualityJobInput:CreateDataQualityJobDefinition', createDataQualityJobDefinition_dataQualityJobInput - A list of inputs for the monitoring job. Currently endpoints are supported as monitoring inputs.

$sel:dataQualityJobOutputConfig:CreateDataQualityJobDefinition', createDataQualityJobDefinition_dataQualityJobOutputConfig - Undocumented member.

$sel:jobResources:CreateDataQualityJobDefinition', createDataQualityJobDefinition_jobResources - Undocumented member.

$sel:roleArn:CreateDataQualityJobDefinition', createDataQualityJobDefinition_roleArn - The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

Request Lenses

createDataQualityJobDefinition_networkConfig :: Lens' CreateDataQualityJobDefinition (Maybe MonitoringNetworkConfig)

Specifies networking configuration for the monitoring job.

createDataQualityJobDefinition_dataQualityBaselineConfig :: Lens' CreateDataQualityJobDefinition (Maybe DataQualityBaselineConfig)

Configures the constraints and baselines for the monitoring job.

createDataQualityJobDefinition_tags :: Lens' CreateDataQualityJobDefinition (Maybe [Tag])

(Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

createDataQualityJobDefinition_stoppingCondition :: Lens' CreateDataQualityJobDefinition (Maybe MonitoringStoppingCondition)

Undocumented member.

createDataQualityJobDefinition_jobDefinitionName :: Lens' CreateDataQualityJobDefinition Text

The name for the monitoring job definition.

createDataQualityJobDefinition_dataQualityAppSpecification :: Lens' CreateDataQualityJobDefinition DataQualityAppSpecification

Specifies the container that runs the monitoring job.

createDataQualityJobDefinition_dataQualityJobInput :: Lens' CreateDataQualityJobDefinition DataQualityJobInput

A list of inputs for the monitoring job. Currently endpoints are supported as monitoring inputs.

createDataQualityJobDefinition_dataQualityJobOutputConfig :: Lens' CreateDataQualityJobDefinition MonitoringOutputConfig

Undocumented member.

createDataQualityJobDefinition_jobResources :: Lens' CreateDataQualityJobDefinition MonitoringResources

Undocumented member.

createDataQualityJobDefinition_roleArn :: Lens' CreateDataQualityJobDefinition Text

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

Destructuring the Response

data CreateDataQualityJobDefinitionResponse

See: newCreateDataQualityJobDefinitionResponse smart constructor.

Constructors

CreateDataQualityJobDefinitionResponse'

Fields

httpStatus :: Int

The response's http status code.

jobDefinitionArn :: Text

The Amazon Resource Name (ARN) of the job definition.

newCreateDataQualityJobDefinitionResponse

Arguments

:: Int
$sel:httpStatus:CreateDataQualityJobDefinitionResponse'

-> Text
$sel:jobDefinitionArn:CreateDataQualityJobDefinitionResponse'

-> CreateDataQualityJobDefinitionResponse

Create a value of CreateDataQualityJobDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateDataQualityJobDefinitionResponse', createDataQualityJobDefinitionResponse_httpStatus - The response's http status code.

$sel:jobDefinitionArn:CreateDataQualityJobDefinitionResponse', createDataQualityJobDefinitionResponse_jobDefinitionArn - The Amazon Resource Name (ARN) of the job definition.

Response Lenses

createDataQualityJobDefinitionResponse_httpStatus :: Lens' CreateDataQualityJobDefinitionResponse Int

The response's http status code.

createDataQualityJobDefinitionResponse_jobDefinitionArn :: Lens' CreateDataQualityJobDefinitionResponse Text

The Amazon Resource Name (ARN) of the job definition.