Network.AWS.SageMaker.CreateEdgePackagingJob

Description

Starts a SageMaker Edge Manager model packaging job. Edge Manager will use the model artifacts from the Amazon Simple Storage Service bucket that you specify. After the model has been packaged, Amazon SageMaker saves the resulting artifacts to an S3 bucket that you specify.

Creating a Request

data CreateEdgePackagingJob

See: newCreateEdgePackagingJob smart constructor.

Constructors

CreateEdgePackagingJob'

Fields

resourceKey :: Maybe Text

The Amazon Web Services KMS key to use when encrypting the EBS volume the edge packaging job runs on.

tags :: Maybe [Tag]

Creates tags for the packaging job.

edgePackagingJobName :: Text

The name of the edge packaging job.

compilationJobName :: Text

The name of the SageMaker Neo compilation job that will be used to locate model artifacts for packaging.

modelName :: Text

The name of the model.

modelVersion :: Text

The version of the model.

roleArn :: Text

The Amazon Resource Name (ARN) of an IAM role that enables Amazon SageMaker to download and upload the model, and to contact SageMaker Neo.

outputConfig :: EdgeOutputConfig

Provides information about the output location for the packaged model.

newCreateEdgePackagingJob

Arguments

:: Text
$sel:edgePackagingJobName:CreateEdgePackagingJob'

-> Text
$sel:compilationJobName:CreateEdgePackagingJob'

-> Text
$sel:modelName:CreateEdgePackagingJob'

-> Text
$sel:modelVersion:CreateEdgePackagingJob'

-> Text
$sel:roleArn:CreateEdgePackagingJob'

-> EdgeOutputConfig
$sel:outputConfig:CreateEdgePackagingJob'

-> CreateEdgePackagingJob

Create a value of CreateEdgePackagingJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceKey:CreateEdgePackagingJob', createEdgePackagingJob_resourceKey - The Amazon Web Services KMS key to use when encrypting the EBS volume the edge packaging job runs on.

$sel:tags:CreateEdgePackagingJob', createEdgePackagingJob_tags - Creates tags for the packaging job.

$sel:edgePackagingJobName:CreateEdgePackagingJob', createEdgePackagingJob_edgePackagingJobName - The name of the edge packaging job.

$sel:compilationJobName:CreateEdgePackagingJob', createEdgePackagingJob_compilationJobName - The name of the SageMaker Neo compilation job that will be used to locate model artifacts for packaging.

$sel:modelName:CreateEdgePackagingJob', createEdgePackagingJob_modelName - The name of the model.

$sel:modelVersion:CreateEdgePackagingJob', createEdgePackagingJob_modelVersion - The version of the model.

$sel:roleArn:CreateEdgePackagingJob', createEdgePackagingJob_roleArn - The Amazon Resource Name (ARN) of an IAM role that enables Amazon SageMaker to download and upload the model, and to contact SageMaker Neo.

$sel:outputConfig:CreateEdgePackagingJob', createEdgePackagingJob_outputConfig - Provides information about the output location for the packaged model.

Request Lenses

createEdgePackagingJob_resourceKey :: Lens' CreateEdgePackagingJob (Maybe Text)

The Amazon Web Services KMS key to use when encrypting the EBS volume the edge packaging job runs on.

createEdgePackagingJob_tags :: Lens' CreateEdgePackagingJob (Maybe [Tag])

Creates tags for the packaging job.

createEdgePackagingJob_edgePackagingJobName :: Lens' CreateEdgePackagingJob Text

The name of the edge packaging job.

createEdgePackagingJob_compilationJobName :: Lens' CreateEdgePackagingJob Text

The name of the SageMaker Neo compilation job that will be used to locate model artifacts for packaging.

createEdgePackagingJob_modelName :: Lens' CreateEdgePackagingJob Text

The name of the model.

createEdgePackagingJob_modelVersion :: Lens' CreateEdgePackagingJob Text

The version of the model.

createEdgePackagingJob_roleArn :: Lens' CreateEdgePackagingJob Text

The Amazon Resource Name (ARN) of an IAM role that enables Amazon SageMaker to download and upload the model, and to contact SageMaker Neo.

createEdgePackagingJob_outputConfig :: Lens' CreateEdgePackagingJob EdgeOutputConfig

Provides information about the output location for the packaged model.

Destructuring the Response

data CreateEdgePackagingJobResponse

See: newCreateEdgePackagingJobResponse smart constructor.

Constructors

CreateEdgePackagingJobResponse'

newCreateEdgePackagingJobResponse :: CreateEdgePackagingJobResponse

Create a value of CreateEdgePackagingJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.