Network.AWS.SageMaker.CreateEndpoint

Description

Creates an endpoint using the endpoint configuration specified in the request. Amazon SageMaker uses the endpoint to provision resources and deploy models. You create the endpoint configuration with the CreateEndpointConfig API.

Use this API to deploy models using Amazon SageMaker hosting services.

For an example that calls this method when deploying a model to Amazon SageMaker hosting services, see the Create Endpoint example notebook.

You must not delete an EndpointConfig that is in use by an endpoint that is live or while the UpdateEndpoint or CreateEndpoint operations are being performed on the endpoint. To update an endpoint, you must create a new EndpointConfig.

The endpoint name must be unique within an Amazon Web Services Region in your Amazon Web Services account.

When it receives the request, Amazon SageMaker creates the endpoint, launches the resources (ML compute instances), and deploys the model(s) on them.

When you call CreateEndpoint, a load call is made to DynamoDB to verify that your endpoint configuration exists. When you read data from a DynamoDB table supporting Eventually Consistent Reads , the response might not reflect the results of a recently completed write operation. The response might include some stale data. If the dependent entities are not yet in DynamoDB, this causes a validation error. If you repeat your read request after a short time, the response should return the latest data. So retry logic is recommended to handle these possible issues. We also recommend that customers call DescribeEndpointConfig before calling CreateEndpoint to minimize the potential impact of a DynamoDB eventually consistent read.

When Amazon SageMaker receives the request, it sets the endpoint status to Creating. After it creates the endpoint, it sets the status to InService. Amazon SageMaker can then process incoming requests for inferences. To check the status of an endpoint, use the DescribeEndpoint API.

If any of the models hosted at this endpoint get model data from an Amazon S3 location, Amazon SageMaker uses Amazon Web Services Security Token Service to download model artifacts from the S3 path you provided. Amazon Web Services STS is activated in your IAM user account by default. If you previously deactivated Amazon Web Services STS for a region, you need to reactivate Amazon Web Services STS for that region. For more information, see Activating and Deactivating Amazon Web Services STS in an Amazon Web Services Region in the Amazon Web Services Identity and Access Management User Guide.

To add the IAM role policies for using this API operation, go to the IAM console, and choose Roles in the left navigation pane. Search the IAM role that you want to grant access to use the CreateEndpoint and CreateEndpointConfig API operations, add the following policies to the role.

Option 1: For a full Amazon SageMaker access, search and attach the AmazonSageMakerFullAccess policy.

Option 2: For granting a limited access to an IAM role, paste the following Action elements manually into the JSON file of the IAM role:

"Action": ["sagemaker:CreateEndpoint", "sagemaker:CreateEndpointConfig"]

"Resource": [

"arn:aws:sagemaker:region:account-id:endpoint/endpointName"

"arn:aws:sagemaker:region:account-id:endpoint-config/endpointConfigName"

]

For more information, see Amazon SageMaker API Permissions: Actions, Permissions, and Resources Reference.

Creating a Request

data CreateEndpoint

See: newCreateEndpoint smart constructor.

Constructors

CreateEndpoint'

Fields

tags :: Maybe [Tag]

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

endpointName :: Text

The name of the endpoint.The name must be unique within an Amazon Web Services Region in your Amazon Web Services account. The name is case-insensitive in CreateEndpoint, but the case is preserved and must be matched in .

endpointConfigName :: Text

The name of an endpoint configuration. For more information, see CreateEndpointConfig.

newCreateEndpoint

Arguments

:: Text
$sel:endpointName:CreateEndpoint'

-> Text
$sel:endpointConfigName:CreateEndpoint'

-> CreateEndpoint

Create a value of CreateEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateEndpoint', createEndpoint_tags - An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

$sel:endpointName:CreateEndpoint', createEndpoint_endpointName - The name of the endpoint.The name must be unique within an Amazon Web Services Region in your Amazon Web Services account. The name is case-insensitive in CreateEndpoint, but the case is preserved and must be matched in .

$sel:endpointConfigName:CreateEndpoint', createEndpoint_endpointConfigName - The name of an endpoint configuration. For more information, see CreateEndpointConfig.

Request Lenses

createEndpoint_tags :: Lens' CreateEndpoint (Maybe [Tag])

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

createEndpoint_endpointName :: Lens' CreateEndpoint Text

The name of the endpoint.The name must be unique within an Amazon Web Services Region in your Amazon Web Services account. The name is case-insensitive in CreateEndpoint, but the case is preserved and must be matched in .

createEndpoint_endpointConfigName :: Lens' CreateEndpoint Text

The name of an endpoint configuration. For more information, see CreateEndpointConfig.

Destructuring the Response

data CreateEndpointResponse

See: newCreateEndpointResponse smart constructor.

Constructors

CreateEndpointResponse'

Fields

httpStatus :: Int

The response's http status code.

endpointArn :: Text

The Amazon Resource Name (ARN) of the endpoint.

newCreateEndpointResponse

Arguments

:: Int
$sel:httpStatus:CreateEndpointResponse'

-> Text
$sel:endpointArn:CreateEndpointResponse'

-> CreateEndpointResponse

Create a value of CreateEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateEndpointResponse', createEndpointResponse_httpStatus - The response's http status code.

$sel:endpointArn:CreateEndpointResponse', createEndpointResponse_endpointArn - The Amazon Resource Name (ARN) of the endpoint.

Response Lenses

createEndpointResponse_httpStatus :: Lens' CreateEndpointResponse Int

The response's http status code.

createEndpointResponse_endpointArn :: Lens' CreateEndpointResponse Text

The Amazon Resource Name (ARN) of the endpoint.