Network.AWS.SageMaker.CreateEndpointConfig

Description

Creates an endpoint configuration that Amazon SageMaker hosting services uses to deploy models. In the configuration, you identify one or more models, created using the CreateModel API, to deploy and the resources that you want Amazon SageMaker to provision. Then you call the CreateEndpoint API.

Use this API if you want to use Amazon SageMaker hosting services to deploy models into production.

In the request, you define a ProductionVariant, for each model that you want to deploy. Each ProductionVariant parameter also describes the resources that you want Amazon SageMaker to provision. This includes the number and type of ML compute instances to deploy.

If you are hosting multiple models, you also assign a VariantWeight to specify how much traffic you want to allocate to each model. For example, suppose that you want to host two models, A and B, and you assign traffic weight 2 for model A and 1 for model B. Amazon SageMaker distributes two-thirds of the traffic to Model A, and one-third to model B.

When you call CreateEndpoint, a load call is made to DynamoDB to verify that your endpoint configuration exists. When you read data from a DynamoDB table supporting Eventually Consistent Reads , the response might not reflect the results of a recently completed write operation. The response might include some stale data. If the dependent entities are not yet in DynamoDB, this causes a validation error. If you repeat your read request after a short time, the response should return the latest data. So retry logic is recommended to handle these possible issues. We also recommend that customers call DescribeEndpointConfig before calling CreateEndpoint to minimize the potential impact of a DynamoDB eventually consistent read.

Creating a Request

data CreateEndpointConfig

See: newCreateEndpointConfig smart constructor.

Constructors

CreateEndpointConfig'

Fields

asyncInferenceConfig :: Maybe AsyncInferenceConfig

Specifies configuration for how an endpoint performs asynchronous inference. This is a required field in order for your Endpoint to be invoked using InvokeEndpointAsync .

kmsKeyId :: Maybe Text

The Amazon Resource Name (ARN) of a Amazon Web Services Key Management Service key that Amazon SageMaker uses to encrypt data on the storage volume attached to the ML compute instance that hosts the endpoint.

The KmsKeyId can be any of the following formats:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Alias name: alias/ExampleAlias

Alias name ARN: arn:aws:kms:us-west-2:111122223333:alias/ExampleAlias

The KMS key policy must grant permission to the IAM role that you specify in your CreateEndpoint, UpdateEndpoint requests. For more information, refer to the Amazon Web Services Key Management Service section Using Key Policies in Amazon Web Services KMS

Certain Nitro-based instances include local storage, dependent on the instance type. Local storage volumes are encrypted using a hardware module on the instance. You can't request a KmsKeyId when using an instance type with local storage. If any of the models that you specify in the ProductionVariants parameter use nitro-based instances with local storage, do not specify a value for the KmsKeyId parameter. If you specify a value for KmsKeyId when using any nitro-based instances with local storage, the call to CreateEndpointConfig fails.

For a list of instance types that support local instance storage, see Instance Store Volumes.

For more information about local instance storage encryption, see SSD Instance Store Volumes.

dataCaptureConfig :: Maybe DataCaptureConfig

tags :: Maybe [Tag]

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

endpointConfigName :: Text

The name of the endpoint configuration. You specify this name in a CreateEndpoint request.

productionVariants :: NonEmpty ProductionVariant

An list of ProductionVariant objects, one for each model that you want to host at this endpoint.

newCreateEndpointConfig

Arguments

:: Text
$sel:endpointConfigName:CreateEndpointConfig'

-> NonEmpty ProductionVariant
$sel:productionVariants:CreateEndpointConfig'

-> CreateEndpointConfig

Create a value of CreateEndpointConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:asyncInferenceConfig:CreateEndpointConfig', createEndpointConfig_asyncInferenceConfig - Specifies configuration for how an endpoint performs asynchronous inference. This is a required field in order for your Endpoint to be invoked using InvokeEndpointAsync .

$sel:kmsKeyId:CreateEndpointConfig', createEndpointConfig_kmsKeyId - The Amazon Resource Name (ARN) of a Amazon Web Services Key Management Service key that Amazon SageMaker uses to encrypt data on the storage volume attached to the ML compute instance that hosts the endpoint.

The KmsKeyId can be any of the following formats:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Alias name: alias/ExampleAlias

Alias name ARN: arn:aws:kms:us-west-2:111122223333:alias/ExampleAlias

The KMS key policy must grant permission to the IAM role that you specify in your CreateEndpoint, UpdateEndpoint requests. For more information, refer to the Amazon Web Services Key Management Service section Using Key Policies in Amazon Web Services KMS

Certain Nitro-based instances include local storage, dependent on the instance type. Local storage volumes are encrypted using a hardware module on the instance. You can't request a KmsKeyId when using an instance type with local storage. If any of the models that you specify in the ProductionVariants parameter use nitro-based instances with local storage, do not specify a value for the KmsKeyId parameter. If you specify a value for KmsKeyId when using any nitro-based instances with local storage, the call to CreateEndpointConfig fails.

For a list of instance types that support local instance storage, see Instance Store Volumes.

For more information about local instance storage encryption, see SSD Instance Store Volumes.

$sel:dataCaptureConfig:CreateEndpointConfig', createEndpointConfig_dataCaptureConfig - Undocumented member.

$sel:tags:CreateEndpointConfig', createEndpointConfig_tags - An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

$sel:endpointConfigName:CreateEndpointConfig', createEndpointConfig_endpointConfigName - The name of the endpoint configuration. You specify this name in a CreateEndpoint request.

$sel:productionVariants:CreateEndpointConfig', createEndpointConfig_productionVariants - An list of ProductionVariant objects, one for each model that you want to host at this endpoint.

Request Lenses

createEndpointConfig_asyncInferenceConfig :: Lens' CreateEndpointConfig (Maybe AsyncInferenceConfig)

Specifies configuration for how an endpoint performs asynchronous inference. This is a required field in order for your Endpoint to be invoked using InvokeEndpointAsync .

createEndpointConfig_kmsKeyId :: Lens' CreateEndpointConfig (Maybe Text)

The Amazon Resource Name (ARN) of a Amazon Web Services Key Management Service key that Amazon SageMaker uses to encrypt data on the storage volume attached to the ML compute instance that hosts the endpoint.

The KmsKeyId can be any of the following formats:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Alias name: alias/ExampleAlias

Alias name ARN: arn:aws:kms:us-west-2:111122223333:alias/ExampleAlias

The KMS key policy must grant permission to the IAM role that you specify in your CreateEndpoint, UpdateEndpoint requests. For more information, refer to the Amazon Web Services Key Management Service section Using Key Policies in Amazon Web Services KMS

Certain Nitro-based instances include local storage, dependent on the instance type. Local storage volumes are encrypted using a hardware module on the instance. You can't request a KmsKeyId when using an instance type with local storage. If any of the models that you specify in the ProductionVariants parameter use nitro-based instances with local storage, do not specify a value for the KmsKeyId parameter. If you specify a value for KmsKeyId when using any nitro-based instances with local storage, the call to CreateEndpointConfig fails.

For a list of instance types that support local instance storage, see Instance Store Volumes.

For more information about local instance storage encryption, see SSD Instance Store Volumes.

createEndpointConfig_dataCaptureConfig :: Lens' CreateEndpointConfig (Maybe DataCaptureConfig)

Undocumented member.

createEndpointConfig_tags :: Lens' CreateEndpointConfig (Maybe [Tag])

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

createEndpointConfig_endpointConfigName :: Lens' CreateEndpointConfig Text

The name of the endpoint configuration. You specify this name in a CreateEndpoint request.

createEndpointConfig_productionVariants :: Lens' CreateEndpointConfig (NonEmpty ProductionVariant)

An list of ProductionVariant objects, one for each model that you want to host at this endpoint.

Destructuring the Response

data CreateEndpointConfigResponse

See: newCreateEndpointConfigResponse smart constructor.

Constructors

CreateEndpointConfigResponse'

Fields

httpStatus :: Int

The response's http status code.

endpointConfigArn :: Text

The Amazon Resource Name (ARN) of the endpoint configuration.

newCreateEndpointConfigResponse

Arguments

:: Int
$sel:httpStatus:CreateEndpointConfigResponse'

-> Text
$sel:endpointConfigArn:CreateEndpointConfigResponse'

-> CreateEndpointConfigResponse

Create a value of CreateEndpointConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateEndpointConfigResponse', createEndpointConfigResponse_httpStatus - The response's http status code.

$sel:endpointConfigArn:CreateEndpointConfigResponse', createEndpointConfigResponse_endpointConfigArn - The Amazon Resource Name (ARN) of the endpoint configuration.

Response Lenses

createEndpointConfigResponse_httpStatus :: Lens' CreateEndpointConfigResponse Int

The response's http status code.

createEndpointConfigResponse_endpointConfigArn :: Lens' CreateEndpointConfigResponse Text

The Amazon Resource Name (ARN) of the endpoint configuration.