Network.AWS.SageMaker.CreateExperiment

Description

Creates an SageMaker experiment. An experiment is a collection of trials that are observed, compared and evaluated as a group. A trial is a set of steps, called trial components, that produce a machine learning model.

The goal of an experiment is to determine the components that produce the best model. Multiple trials are performed, each one isolating and measuring the impact of a change to one or more inputs, while keeping the remaining inputs constant.

When you use SageMaker Studio or the SageMaker Python SDK, all experiments, trials, and trial components are automatically tracked, logged, and indexed. When you use the Amazon Web Services SDK for Python (Boto), you must use the logging APIs provided by the SDK.

You can add tags to experiments, trials, trial components and then use the Search API to search for the tags.

To add a description to an experiment, specify the optional Description parameter. To add a description later, or to change the description, call the UpdateExperiment API.

To get a list of all your experiments, call the ListExperiments API. To view an experiment's properties, call the DescribeExperiment API. To get a list of all the trials associated with an experiment, call the ListTrials API. To create a trial call the CreateTrial API.

Creating a Request

data CreateExperiment

See: newCreateExperiment smart constructor.

Constructors

CreateExperiment'

Fields

tags :: Maybe [Tag]

A list of tags to associate with the experiment. You can use Search API to search on the tags.

description :: Maybe Text

The description of the experiment.

displayName :: Maybe Text

The name of the experiment as displayed. The name doesn't need to be unique. If you don't specify DisplayName, the value in ExperimentName is displayed.

experimentName :: Text

The name of the experiment. The name must be unique in your Amazon Web Services account and is not case-sensitive.

newCreateExperiment

Arguments

:: Text
$sel:experimentName:CreateExperiment'

-> CreateExperiment

Create a value of CreateExperiment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateExperiment', createExperiment_tags - A list of tags to associate with the experiment. You can use Search API to search on the tags.

$sel:description:CreateExperiment', createExperiment_description - The description of the experiment.

$sel:displayName:CreateExperiment', createExperiment_displayName - The name of the experiment as displayed. The name doesn't need to be unique. If you don't specify DisplayName, the value in ExperimentName is displayed.

$sel:experimentName:CreateExperiment', createExperiment_experimentName - The name of the experiment. The name must be unique in your Amazon Web Services account and is not case-sensitive.

Request Lenses

createExperiment_tags :: Lens' CreateExperiment (Maybe [Tag])

A list of tags to associate with the experiment. You can use Search API to search on the tags.

createExperiment_description :: Lens' CreateExperiment (Maybe Text)

The description of the experiment.

createExperiment_displayName :: Lens' CreateExperiment (Maybe Text)

The name of the experiment as displayed. The name doesn't need to be unique. If you don't specify DisplayName, the value in ExperimentName is displayed.

createExperiment_experimentName :: Lens' CreateExperiment Text

The name of the experiment. The name must be unique in your Amazon Web Services account and is not case-sensitive.

Destructuring the Response

data CreateExperimentResponse

See: newCreateExperimentResponse smart constructor.

Constructors

CreateExperimentResponse'

Fields

experimentArn :: Maybe Text

The Amazon Resource Name (ARN) of the experiment.

httpStatus :: Int

The response's http status code.

newCreateExperimentResponse

Arguments

:: Int
$sel:httpStatus:CreateExperimentResponse'

-> CreateExperimentResponse

Create a value of CreateExperimentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:experimentArn:CreateExperimentResponse', createExperimentResponse_experimentArn - The Amazon Resource Name (ARN) of the experiment.

$sel:httpStatus:CreateExperimentResponse', createExperimentResponse_httpStatus - The response's http status code.

Response Lenses

createExperimentResponse_experimentArn :: Lens' CreateExperimentResponse (Maybe Text)

The Amazon Resource Name (ARN) of the experiment.

createExperimentResponse_httpStatus :: Lens' CreateExperimentResponse Int

The response's http status code.