Network.AWS.SageMaker.CreateFeatureGroup

Description

Create a new FeatureGroup. A FeatureGroup is a group of Features defined in the FeatureStore to describe a Record.

The FeatureGroup defines the schema and features contained in the FeatureGroup. A FeatureGroup definition is composed of a list of Features, a RecordIdentifierFeatureName, an EventTimeFeatureName and configurations for its OnlineStore and OfflineStore. Check Amazon Web Services service quotas to see the FeatureGroups quota for your Amazon Web Services account.

You must include at least one of OnlineStoreConfig and OfflineStoreConfig to create a FeatureGroup.

Creating a Request

data CreateFeatureGroup

See: newCreateFeatureGroup smart constructor.

Constructors

CreateFeatureGroup'

Fields

offlineStoreConfig :: Maybe OfflineStoreConfig

Use this to configure an OfflineFeatureStore. This parameter allows you to specify:

The Amazon Simple Storage Service (Amazon S3) location of an OfflineStore.

A configuration for an Amazon Web Services Glue or Amazon Web Services Hive data cataolgue.

An KMS encryption key to encrypt the Amazon S3 location used for OfflineStore.

To learn more about this parameter, see OfflineStoreConfig.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM execution role used to persist data into the OfflineStore if an OfflineStoreConfig is provided.

tags :: Maybe [Tag]

Tags used to identify Features in each FeatureGroup.

description :: Maybe Text

A free-form description of a FeatureGroup.

onlineStoreConfig :: Maybe OnlineStoreConfig

You can turn the OnlineStore on or off by specifying True for the EnableOnlineStore flag in OnlineStoreConfig; the default value is False.

You can also include an Amazon Web Services KMS key ID (KMSKeyId) for at-rest encryption of the OnlineStore.

featureGroupName :: Text

The name of the FeatureGroup. The name must be unique within an Amazon Web Services Region in an Amazon Web Services account. The name:

Must start and end with an alphanumeric character.

Can only contain alphanumeric character and hyphens. Spaces are not allowed.

recordIdentifierFeatureName :: Text

The name of the Feature whose value uniquely identifies a Record defined in the FeatureStore. Only the latest record per identifier value will be stored in the OnlineStore. RecordIdentifierFeatureName must be one of feature definitions' names.

You use the RecordIdentifierFeatureName to access data in a FeatureStore.

This name:

Must start and end with an alphanumeric character.

Can only contains alphanumeric characters, hyphens, underscores. Spaces are not allowed.

eventTimeFeatureName :: Text

The name of the feature that stores the EventTime of a Record in a FeatureGroup.

An EventTime is a point in time when a new event occurs that corresponds to the creation or update of a Record in a FeatureGroup. All Records in the FeatureGroup must have a corresponding EventTime.

An EventTime can be a String or Fractional.

Fractional: EventTime feature values must be a Unix timestamp in seconds.

String: EventTime feature values must be an ISO-8601 string in the format. The following formats are supported yyyy-MM-dd'T'HH:mm:ssZ and yyyy-MM-dd'T'HH:mm:ss.SSSZ where yyyy, MM, and dd represent the year, month, and day respectively and HH, mm, ss, and if applicable, SSS represent the hour, month, second and milliseconds respsectively. 'T' and Z are constants.

featureDefinitions :: NonEmpty FeatureDefinition

A list of Feature names and types. Name and Type is compulsory per Feature.

Valid feature FeatureTypes are Integral, Fractional and String.

FeatureNames cannot be any of the following: is_deleted, write_time, api_invocation_time

You can create up to 2,500 FeatureDefinitions per FeatureGroup.

newCreateFeatureGroup

Arguments

:: Text
$sel:featureGroupName:CreateFeatureGroup'

-> Text
$sel:recordIdentifierFeatureName:CreateFeatureGroup'

-> Text
$sel:eventTimeFeatureName:CreateFeatureGroup'

-> NonEmpty FeatureDefinition
$sel:featureDefinitions:CreateFeatureGroup'

-> CreateFeatureGroup

Create a value of CreateFeatureGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:offlineStoreConfig:CreateFeatureGroup', createFeatureGroup_offlineStoreConfig - Use this to configure an OfflineFeatureStore. This parameter allows you to specify:

The Amazon Simple Storage Service (Amazon S3) location of an OfflineStore.

A configuration for an Amazon Web Services Glue or Amazon Web Services Hive data cataolgue.

An KMS encryption key to encrypt the Amazon S3 location used for OfflineStore.

To learn more about this parameter, see OfflineStoreConfig.

$sel:roleArn:CreateFeatureGroup', createFeatureGroup_roleArn - The Amazon Resource Name (ARN) of the IAM execution role used to persist data into the OfflineStore if an OfflineStoreConfig is provided.

$sel:tags:CreateFeatureGroup', createFeatureGroup_tags - Tags used to identify Features in each FeatureGroup.

$sel:description:CreateFeatureGroup', createFeatureGroup_description - A free-form description of a FeatureGroup.

$sel:onlineStoreConfig:CreateFeatureGroup', createFeatureGroup_onlineStoreConfig - You can turn the OnlineStore on or off by specifying True for the EnableOnlineStore flag in OnlineStoreConfig; the default value is False.

You can also include an Amazon Web Services KMS key ID (KMSKeyId) for at-rest encryption of the OnlineStore.

$sel:featureGroupName:CreateFeatureGroup', createFeatureGroup_featureGroupName - The name of the FeatureGroup. The name must be unique within an Amazon Web Services Region in an Amazon Web Services account. The name:

Must start and end with an alphanumeric character.

Can only contain alphanumeric character and hyphens. Spaces are not allowed.

$sel:recordIdentifierFeatureName:CreateFeatureGroup', createFeatureGroup_recordIdentifierFeatureName - The name of the Feature whose value uniquely identifies a Record defined in the FeatureStore. Only the latest record per identifier value will be stored in the OnlineStore. RecordIdentifierFeatureName must be one of feature definitions' names.

You use the RecordIdentifierFeatureName to access data in a FeatureStore.

This name:

Must start and end with an alphanumeric character.

Can only contains alphanumeric characters, hyphens, underscores. Spaces are not allowed.

$sel:eventTimeFeatureName:CreateFeatureGroup', createFeatureGroup_eventTimeFeatureName - The name of the feature that stores the EventTime of a Record in a FeatureGroup.

An EventTime is a point in time when a new event occurs that corresponds to the creation or update of a Record in a FeatureGroup. All Records in the FeatureGroup must have a corresponding EventTime.

An EventTime can be a String or Fractional.

Fractional: EventTime feature values must be a Unix timestamp in seconds.

String: EventTime feature values must be an ISO-8601 string in the format. The following formats are supported yyyy-MM-dd'T'HH:mm:ssZ and yyyy-MM-dd'T'HH:mm:ss.SSSZ where yyyy, MM, and dd represent the year, month, and day respectively and HH, mm, ss, and if applicable, SSS represent the hour, month, second and milliseconds respsectively. 'T' and Z are constants.

$sel:featureDefinitions:CreateFeatureGroup', createFeatureGroup_featureDefinitions - A list of Feature names and types. Name and Type is compulsory per Feature.

Valid feature FeatureTypes are Integral, Fractional and String.

FeatureNames cannot be any of the following: is_deleted, write_time, api_invocation_time

You can create up to 2,500 FeatureDefinitions per FeatureGroup.

Request Lenses

createFeatureGroup_offlineStoreConfig :: Lens' CreateFeatureGroup (Maybe OfflineStoreConfig)

Use this to configure an OfflineFeatureStore. This parameter allows you to specify:

The Amazon Simple Storage Service (Amazon S3) location of an OfflineStore.

A configuration for an Amazon Web Services Glue or Amazon Web Services Hive data cataolgue.

An KMS encryption key to encrypt the Amazon S3 location used for OfflineStore.

To learn more about this parameter, see OfflineStoreConfig.

createFeatureGroup_roleArn :: Lens' CreateFeatureGroup (Maybe Text)

The Amazon Resource Name (ARN) of the IAM execution role used to persist data into the OfflineStore if an OfflineStoreConfig is provided.

createFeatureGroup_tags :: Lens' CreateFeatureGroup (Maybe [Tag])

Tags used to identify Features in each FeatureGroup.

createFeatureGroup_description :: Lens' CreateFeatureGroup (Maybe Text)

A free-form description of a FeatureGroup.

createFeatureGroup_onlineStoreConfig :: Lens' CreateFeatureGroup (Maybe OnlineStoreConfig)

You can turn the OnlineStore on or off by specifying True for the EnableOnlineStore flag in OnlineStoreConfig; the default value is False.

You can also include an Amazon Web Services KMS key ID (KMSKeyId) for at-rest encryption of the OnlineStore.

createFeatureGroup_featureGroupName :: Lens' CreateFeatureGroup Text

The name of the FeatureGroup. The name must be unique within an Amazon Web Services Region in an Amazon Web Services account. The name:

Must start and end with an alphanumeric character.

Can only contain alphanumeric character and hyphens. Spaces are not allowed.

createFeatureGroup_recordIdentifierFeatureName :: Lens' CreateFeatureGroup Text

The name of the Feature whose value uniquely identifies a Record defined in the FeatureStore. Only the latest record per identifier value will be stored in the OnlineStore. RecordIdentifierFeatureName must be one of feature definitions' names.

You use the RecordIdentifierFeatureName to access data in a FeatureStore.

This name:

Must start and end with an alphanumeric character.

Can only contains alphanumeric characters, hyphens, underscores. Spaces are not allowed.

createFeatureGroup_eventTimeFeatureName :: Lens' CreateFeatureGroup Text

The name of the feature that stores the EventTime of a Record in a FeatureGroup.

An EventTime is a point in time when a new event occurs that corresponds to the creation or update of a Record in a FeatureGroup. All Records in the FeatureGroup must have a corresponding EventTime.

An EventTime can be a String or Fractional.

Fractional: EventTime feature values must be a Unix timestamp in seconds.

String: EventTime feature values must be an ISO-8601 string in the format. The following formats are supported yyyy-MM-dd'T'HH:mm:ssZ and yyyy-MM-dd'T'HH:mm:ss.SSSZ where yyyy, MM, and dd represent the year, month, and day respectively and HH, mm, ss, and if applicable, SSS represent the hour, month, second and milliseconds respsectively. 'T' and Z are constants.

createFeatureGroup_featureDefinitions :: Lens' CreateFeatureGroup (NonEmpty FeatureDefinition)

A list of Feature names and types. Name and Type is compulsory per Feature.

Valid feature FeatureTypes are Integral, Fractional and String.

FeatureNames cannot be any of the following: is_deleted, write_time, api_invocation_time

You can create up to 2,500 FeatureDefinitions per FeatureGroup.

Destructuring the Response

data CreateFeatureGroupResponse

See: newCreateFeatureGroupResponse smart constructor.

Constructors

CreateFeatureGroupResponse'

Fields

httpStatus :: Int

The response's http status code.

featureGroupArn :: Text

The Amazon Resource Name (ARN) of the FeatureGroup. This is a unique identifier for the feature group.

newCreateFeatureGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateFeatureGroupResponse'

-> Text
$sel:featureGroupArn:CreateFeatureGroupResponse'

-> CreateFeatureGroupResponse

Create a value of CreateFeatureGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateFeatureGroupResponse', createFeatureGroupResponse_httpStatus - The response's http status code.

$sel:featureGroupArn:CreateFeatureGroupResponse', createFeatureGroupResponse_featureGroupArn - The Amazon Resource Name (ARN) of the FeatureGroup. This is a unique identifier for the feature group.

Response Lenses

createFeatureGroupResponse_httpStatus :: Lens' CreateFeatureGroupResponse Int

The response's http status code.

createFeatureGroupResponse_featureGroupArn :: Lens' CreateFeatureGroupResponse Text

The Amazon Resource Name (ARN) of the FeatureGroup. This is a unique identifier for the feature group.