Network.AWS.SageMaker.CreateFlowDefinition

Description

Creates a flow definition.

Creating a Request

data CreateFlowDefinition

See: newCreateFlowDefinition smart constructor.

Constructors

CreateFlowDefinition'

Fields

humanLoopRequestSource :: Maybe HumanLoopRequestSource

Container for configuring the source of human task requests. Use to specify if Amazon Rekognition or Amazon Textract is used as an integration source.

humanLoopActivationConfig :: Maybe HumanLoopActivationConfig

An object containing information about the events that trigger a human workflow.

tags :: Maybe [Tag]

An array of key-value pairs that contain metadata to help you categorize and organize a flow definition. Each tag consists of a key and a value, both of which you define.

flowDefinitionName :: Text

The name of your flow definition.

humanLoopConfig :: HumanLoopConfig

An object containing information about the tasks the human reviewers will perform.

outputConfig :: FlowDefinitionOutputConfig

An object containing information about where the human review results will be uploaded.

roleArn :: Text

The Amazon Resource Name (ARN) of the role needed to call other services on your behalf. For example, arn:aws:iam::1234567890:role/service-role/AmazonSageMaker-ExecutionRole-20180111T151298.

newCreateFlowDefinition

Arguments

:: Text
$sel:flowDefinitionName:CreateFlowDefinition'

-> HumanLoopConfig
$sel:humanLoopConfig:CreateFlowDefinition'

-> FlowDefinitionOutputConfig
$sel:outputConfig:CreateFlowDefinition'

-> Text
$sel:roleArn:CreateFlowDefinition'

-> CreateFlowDefinition

Create a value of CreateFlowDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:humanLoopRequestSource:CreateFlowDefinition', createFlowDefinition_humanLoopRequestSource - Container for configuring the source of human task requests. Use to specify if Amazon Rekognition or Amazon Textract is used as an integration source.

$sel:humanLoopActivationConfig:CreateFlowDefinition', createFlowDefinition_humanLoopActivationConfig - An object containing information about the events that trigger a human workflow.

$sel:tags:CreateFlowDefinition', createFlowDefinition_tags - An array of key-value pairs that contain metadata to help you categorize and organize a flow definition. Each tag consists of a key and a value, both of which you define.

$sel:flowDefinitionName:CreateFlowDefinition', createFlowDefinition_flowDefinitionName - The name of your flow definition.

$sel:humanLoopConfig:CreateFlowDefinition', createFlowDefinition_humanLoopConfig - An object containing information about the tasks the human reviewers will perform.

$sel:outputConfig:CreateFlowDefinition', createFlowDefinition_outputConfig - An object containing information about where the human review results will be uploaded.

$sel:roleArn:CreateFlowDefinition', createFlowDefinition_roleArn - The Amazon Resource Name (ARN) of the role needed to call other services on your behalf. For example, arn:aws:iam::1234567890:role/service-role/AmazonSageMaker-ExecutionRole-20180111T151298.

Request Lenses

createFlowDefinition_humanLoopRequestSource :: Lens' CreateFlowDefinition (Maybe HumanLoopRequestSource)

Container for configuring the source of human task requests. Use to specify if Amazon Rekognition or Amazon Textract is used as an integration source.

createFlowDefinition_humanLoopActivationConfig :: Lens' CreateFlowDefinition (Maybe HumanLoopActivationConfig)

An object containing information about the events that trigger a human workflow.

createFlowDefinition_tags :: Lens' CreateFlowDefinition (Maybe [Tag])

An array of key-value pairs that contain metadata to help you categorize and organize a flow definition. Each tag consists of a key and a value, both of which you define.

createFlowDefinition_flowDefinitionName :: Lens' CreateFlowDefinition Text

The name of your flow definition.

createFlowDefinition_humanLoopConfig :: Lens' CreateFlowDefinition HumanLoopConfig

An object containing information about the tasks the human reviewers will perform.

createFlowDefinition_outputConfig :: Lens' CreateFlowDefinition FlowDefinitionOutputConfig

An object containing information about where the human review results will be uploaded.

createFlowDefinition_roleArn :: Lens' CreateFlowDefinition Text

The Amazon Resource Name (ARN) of the role needed to call other services on your behalf. For example, arn:aws:iam::1234567890:role/service-role/AmazonSageMaker-ExecutionRole-20180111T151298.

Destructuring the Response

data CreateFlowDefinitionResponse

See: newCreateFlowDefinitionResponse smart constructor.

Constructors

CreateFlowDefinitionResponse'

Fields

httpStatus :: Int

The response's http status code.

flowDefinitionArn :: Text

The Amazon Resource Name (ARN) of the flow definition you create.

newCreateFlowDefinitionResponse

Arguments

:: Int
$sel:httpStatus:CreateFlowDefinitionResponse'

-> Text
$sel:flowDefinitionArn:CreateFlowDefinitionResponse'

-> CreateFlowDefinitionResponse

Create a value of CreateFlowDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateFlowDefinitionResponse', createFlowDefinitionResponse_httpStatus - The response's http status code.

$sel:flowDefinitionArn:CreateFlowDefinitionResponse', createFlowDefinitionResponse_flowDefinitionArn - The Amazon Resource Name (ARN) of the flow definition you create.

Response Lenses

createFlowDefinitionResponse_httpStatus :: Lens' CreateFlowDefinitionResponse Int

The response's http status code.

createFlowDefinitionResponse_flowDefinitionArn :: Lens' CreateFlowDefinitionResponse Text

The Amazon Resource Name (ARN) of the flow definition you create.