Network.AWS.SageMaker.CreateHyperParameterTuningJob

Description

Starts a hyperparameter tuning job. A hyperparameter tuning job finds the best version of a model by running many training jobs on your dataset using the algorithm you choose and values for hyperparameters within ranges that you specify. It then chooses the hyperparameter values that result in a model that performs the best, as measured by an objective metric that you choose.

Creating a Request

data CreateHyperParameterTuningJob

See: newCreateHyperParameterTuningJob smart constructor.

Constructors

CreateHyperParameterTuningJob'

Fields

trainingJobDefinition :: Maybe HyperParameterTrainingJobDefinition

The HyperParameterTrainingJobDefinition object that describes the training jobs that this tuning job launches, including static hyperparameters, input data configuration, output data configuration, resource configuration, and stopping condition.

warmStartConfig :: Maybe HyperParameterTuningJobWarmStartConfig

Specifies the configuration for starting the hyperparameter tuning job using one or more previous tuning jobs as a starting point. The results of previous tuning jobs are used to inform which combinations of hyperparameters to search over in the new tuning job.

All training jobs launched by the new hyperparameter tuning job are evaluated by using the objective metric. If you specify IDENTICAL_DATA_AND_ALGORITHM as the WarmStartType value for the warm start configuration, the training job that performs the best in the new tuning job is compared to the best training jobs from the parent tuning jobs. From these, the training job that performs the best as measured by the objective metric is returned as the overall best training job.

All training jobs launched by parent hyperparameter tuning jobs and the new hyperparameter tuning jobs count against the limit of training jobs for the tuning job.

tags :: Maybe [Tag]

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

Tags that you specify for the tuning job are also added to all training jobs that the tuning job launches.

trainingJobDefinitions :: Maybe (NonEmpty HyperParameterTrainingJobDefinition)

A list of the HyperParameterTrainingJobDefinition objects launched for this tuning job.

hyperParameterTuningJobName :: Text

The name of the tuning job. This name is the prefix for the names of all training jobs that this tuning job launches. The name must be unique within the same Amazon Web Services account and Amazon Web Services Region. The name must have 1 to 32 characters. Valid characters are a-z, A-Z, 0-9, and : + = @ _ % - (hyphen). The name is not case sensitive.

hyperParameterTuningJobConfig :: HyperParameterTuningJobConfig

The HyperParameterTuningJobConfig object that describes the tuning job, including the search strategy, the objective metric used to evaluate training jobs, ranges of parameters to search, and resource limits for the tuning job. For more information, see How Hyperparameter Tuning Works.

newCreateHyperParameterTuningJob

Arguments

:: Text
$sel:hyperParameterTuningJobName:CreateHyperParameterTuningJob'

-> HyperParameterTuningJobConfig
$sel:hyperParameterTuningJobConfig:CreateHyperParameterTuningJob'

-> CreateHyperParameterTuningJob

Create a value of CreateHyperParameterTuningJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trainingJobDefinition:CreateHyperParameterTuningJob', createHyperParameterTuningJob_trainingJobDefinition - The HyperParameterTrainingJobDefinition object that describes the training jobs that this tuning job launches, including static hyperparameters, input data configuration, output data configuration, resource configuration, and stopping condition.

$sel:warmStartConfig:CreateHyperParameterTuningJob', createHyperParameterTuningJob_warmStartConfig - Specifies the configuration for starting the hyperparameter tuning job using one or more previous tuning jobs as a starting point. The results of previous tuning jobs are used to inform which combinations of hyperparameters to search over in the new tuning job.

All training jobs launched by the new hyperparameter tuning job are evaluated by using the objective metric. If you specify IDENTICAL_DATA_AND_ALGORITHM as the WarmStartType value for the warm start configuration, the training job that performs the best in the new tuning job is compared to the best training jobs from the parent tuning jobs. From these, the training job that performs the best as measured by the objective metric is returned as the overall best training job.

All training jobs launched by parent hyperparameter tuning jobs and the new hyperparameter tuning jobs count against the limit of training jobs for the tuning job.

$sel:tags:CreateHyperParameterTuningJob', createHyperParameterTuningJob_tags - An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

Tags that you specify for the tuning job are also added to all training jobs that the tuning job launches.

$sel:trainingJobDefinitions:CreateHyperParameterTuningJob', createHyperParameterTuningJob_trainingJobDefinitions - A list of the HyperParameterTrainingJobDefinition objects launched for this tuning job.

$sel:hyperParameterTuningJobName:CreateHyperParameterTuningJob', createHyperParameterTuningJob_hyperParameterTuningJobName - The name of the tuning job. This name is the prefix for the names of all training jobs that this tuning job launches. The name must be unique within the same Amazon Web Services account and Amazon Web Services Region. The name must have 1 to 32 characters. Valid characters are a-z, A-Z, 0-9, and : + = @ _ % - (hyphen). The name is not case sensitive.

$sel:hyperParameterTuningJobConfig:CreateHyperParameterTuningJob', createHyperParameterTuningJob_hyperParameterTuningJobConfig - The HyperParameterTuningJobConfig object that describes the tuning job, including the search strategy, the objective metric used to evaluate training jobs, ranges of parameters to search, and resource limits for the tuning job. For more information, see How Hyperparameter Tuning Works.

Request Lenses

createHyperParameterTuningJob_trainingJobDefinition :: Lens' CreateHyperParameterTuningJob (Maybe HyperParameterTrainingJobDefinition)

The HyperParameterTrainingJobDefinition object that describes the training jobs that this tuning job launches, including static hyperparameters, input data configuration, output data configuration, resource configuration, and stopping condition.

createHyperParameterTuningJob_warmStartConfig :: Lens' CreateHyperParameterTuningJob (Maybe HyperParameterTuningJobWarmStartConfig)

Specifies the configuration for starting the hyperparameter tuning job using one or more previous tuning jobs as a starting point. The results of previous tuning jobs are used to inform which combinations of hyperparameters to search over in the new tuning job.

All training jobs launched by the new hyperparameter tuning job are evaluated by using the objective metric. If you specify IDENTICAL_DATA_AND_ALGORITHM as the WarmStartType value for the warm start configuration, the training job that performs the best in the new tuning job is compared to the best training jobs from the parent tuning jobs. From these, the training job that performs the best as measured by the objective metric is returned as the overall best training job.

All training jobs launched by parent hyperparameter tuning jobs and the new hyperparameter tuning jobs count against the limit of training jobs for the tuning job.

createHyperParameterTuningJob_tags :: Lens' CreateHyperParameterTuningJob (Maybe [Tag])

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

Tags that you specify for the tuning job are also added to all training jobs that the tuning job launches.

createHyperParameterTuningJob_trainingJobDefinitions :: Lens' CreateHyperParameterTuningJob (Maybe (NonEmpty HyperParameterTrainingJobDefinition))

A list of the HyperParameterTrainingJobDefinition objects launched for this tuning job.

createHyperParameterTuningJob_hyperParameterTuningJobName :: Lens' CreateHyperParameterTuningJob Text

The name of the tuning job. This name is the prefix for the names of all training jobs that this tuning job launches. The name must be unique within the same Amazon Web Services account and Amazon Web Services Region. The name must have 1 to 32 characters. Valid characters are a-z, A-Z, 0-9, and : + = @ _ % - (hyphen). The name is not case sensitive.

createHyperParameterTuningJob_hyperParameterTuningJobConfig :: Lens' CreateHyperParameterTuningJob HyperParameterTuningJobConfig

The HyperParameterTuningJobConfig object that describes the tuning job, including the search strategy, the objective metric used to evaluate training jobs, ranges of parameters to search, and resource limits for the tuning job. For more information, see How Hyperparameter Tuning Works.

Destructuring the Response

data CreateHyperParameterTuningJobResponse

See: newCreateHyperParameterTuningJobResponse smart constructor.

Constructors

CreateHyperParameterTuningJobResponse'

Fields

httpStatus :: Int

The response's http status code.

hyperParameterTuningJobArn :: Text

The Amazon Resource Name (ARN) of the tuning job. Amazon SageMaker assigns an ARN to a hyperparameter tuning job when you create it.

newCreateHyperParameterTuningJobResponse

Arguments

:: Int
$sel:httpStatus:CreateHyperParameterTuningJobResponse'

-> Text
$sel:hyperParameterTuningJobArn:CreateHyperParameterTuningJobResponse'

-> CreateHyperParameterTuningJobResponse

Create a value of CreateHyperParameterTuningJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateHyperParameterTuningJobResponse', createHyperParameterTuningJobResponse_httpStatus - The response's http status code.

$sel:hyperParameterTuningJobArn:CreateHyperParameterTuningJobResponse', createHyperParameterTuningJobResponse_hyperParameterTuningJobArn - The Amazon Resource Name (ARN) of the tuning job. Amazon SageMaker assigns an ARN to a hyperparameter tuning job when you create it.

Response Lenses

createHyperParameterTuningJobResponse_httpStatus :: Lens' CreateHyperParameterTuningJobResponse Int

The response's http status code.

createHyperParameterTuningJobResponse_hyperParameterTuningJobArn :: Lens' CreateHyperParameterTuningJobResponse Text

The Amazon Resource Name (ARN) of the tuning job. Amazon SageMaker assigns an ARN to a hyperparameter tuning job when you create it.