Network.AWS.SageMaker.CreateModelBiasJobDefinition

Description

Creates the definition for a model bias job.

Creating a Request

data CreateModelBiasJobDefinition

See: newCreateModelBiasJobDefinition smart constructor.

Constructors

CreateModelBiasJobDefinition'

Fields

networkConfig :: Maybe MonitoringNetworkConfig

Networking options for a model bias job.

modelBiasBaselineConfig :: Maybe ModelBiasBaselineConfig

The baseline configuration for a model bias job.

tags :: Maybe [Tag]

(Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

stoppingCondition :: Maybe MonitoringStoppingCondition

jobDefinitionName :: Text

The name of the bias job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

modelBiasAppSpecification :: ModelBiasAppSpecification

Configures the model bias job to run a specified Docker container image.

modelBiasJobInput :: ModelBiasJobInput

Inputs for the model bias job.

modelBiasJobOutputConfig :: MonitoringOutputConfig

jobResources :: MonitoringResources

roleArn :: Text

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

newCreateModelBiasJobDefinition

Arguments

:: Text
$sel:jobDefinitionName:CreateModelBiasJobDefinition'

-> ModelBiasAppSpecification
$sel:modelBiasAppSpecification:CreateModelBiasJobDefinition'

-> ModelBiasJobInput
$sel:modelBiasJobInput:CreateModelBiasJobDefinition'

-> MonitoringOutputConfig
$sel:modelBiasJobOutputConfig:CreateModelBiasJobDefinition'

-> MonitoringResources
$sel:jobResources:CreateModelBiasJobDefinition'

-> Text
$sel:roleArn:CreateModelBiasJobDefinition'

-> CreateModelBiasJobDefinition

Create a value of CreateModelBiasJobDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkConfig:CreateModelBiasJobDefinition', createModelBiasJobDefinition_networkConfig - Networking options for a model bias job.

$sel:modelBiasBaselineConfig:CreateModelBiasJobDefinition', createModelBiasJobDefinition_modelBiasBaselineConfig - The baseline configuration for a model bias job.

$sel:tags:CreateModelBiasJobDefinition', createModelBiasJobDefinition_tags - (Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

$sel:stoppingCondition:CreateModelBiasJobDefinition', createModelBiasJobDefinition_stoppingCondition - Undocumented member.

$sel:jobDefinitionName:CreateModelBiasJobDefinition', createModelBiasJobDefinition_jobDefinitionName - The name of the bias job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

$sel:modelBiasAppSpecification:CreateModelBiasJobDefinition', createModelBiasJobDefinition_modelBiasAppSpecification - Configures the model bias job to run a specified Docker container image.

$sel:modelBiasJobInput:CreateModelBiasJobDefinition', createModelBiasJobDefinition_modelBiasJobInput - Inputs for the model bias job.

$sel:modelBiasJobOutputConfig:CreateModelBiasJobDefinition', createModelBiasJobDefinition_modelBiasJobOutputConfig - Undocumented member.

$sel:jobResources:CreateModelBiasJobDefinition', createModelBiasJobDefinition_jobResources - Undocumented member.

$sel:roleArn:CreateModelBiasJobDefinition', createModelBiasJobDefinition_roleArn - The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

Request Lenses

createModelBiasJobDefinition_networkConfig :: Lens' CreateModelBiasJobDefinition (Maybe MonitoringNetworkConfig)

Networking options for a model bias job.

createModelBiasJobDefinition_modelBiasBaselineConfig :: Lens' CreateModelBiasJobDefinition (Maybe ModelBiasBaselineConfig)

The baseline configuration for a model bias job.

createModelBiasJobDefinition_tags :: Lens' CreateModelBiasJobDefinition (Maybe [Tag])

(Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

createModelBiasJobDefinition_stoppingCondition :: Lens' CreateModelBiasJobDefinition (Maybe MonitoringStoppingCondition)

Undocumented member.

createModelBiasJobDefinition_jobDefinitionName :: Lens' CreateModelBiasJobDefinition Text

The name of the bias job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

createModelBiasJobDefinition_modelBiasAppSpecification :: Lens' CreateModelBiasJobDefinition ModelBiasAppSpecification

Configures the model bias job to run a specified Docker container image.

createModelBiasJobDefinition_modelBiasJobInput :: Lens' CreateModelBiasJobDefinition ModelBiasJobInput

Inputs for the model bias job.

createModelBiasJobDefinition_modelBiasJobOutputConfig :: Lens' CreateModelBiasJobDefinition MonitoringOutputConfig

Undocumented member.

createModelBiasJobDefinition_jobResources :: Lens' CreateModelBiasJobDefinition MonitoringResources

Undocumented member.

createModelBiasJobDefinition_roleArn :: Lens' CreateModelBiasJobDefinition Text

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

Destructuring the Response

data CreateModelBiasJobDefinitionResponse

See: newCreateModelBiasJobDefinitionResponse smart constructor.

Constructors

CreateModelBiasJobDefinitionResponse'

Fields

httpStatus :: Int

The response's http status code.

jobDefinitionArn :: Text

The Amazon Resource Name (ARN) of the model bias job.

newCreateModelBiasJobDefinitionResponse

Arguments

:: Int
$sel:httpStatus:CreateModelBiasJobDefinitionResponse'

-> Text
$sel:jobDefinitionArn:CreateModelBiasJobDefinitionResponse'

-> CreateModelBiasJobDefinitionResponse

Create a value of CreateModelBiasJobDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateModelBiasJobDefinitionResponse', createModelBiasJobDefinitionResponse_httpStatus - The response's http status code.

$sel:jobDefinitionArn:CreateModelBiasJobDefinitionResponse', createModelBiasJobDefinitionResponse_jobDefinitionArn - The Amazon Resource Name (ARN) of the model bias job.

Response Lenses

createModelBiasJobDefinitionResponse_httpStatus :: Lens' CreateModelBiasJobDefinitionResponse Int

The response's http status code.

createModelBiasJobDefinitionResponse_jobDefinitionArn :: Lens' CreateModelBiasJobDefinitionResponse Text

The Amazon Resource Name (ARN) of the model bias job.