Network.AWS.SageMaker.CreateModelExplainabilityJobDefinition

Description

Creates the definition for a model explainability job.

Creating a Request

data CreateModelExplainabilityJobDefinition

See: newCreateModelExplainabilityJobDefinition smart constructor.

Constructors

CreateModelExplainabilityJobDefinition'

Fields

networkConfig :: Maybe MonitoringNetworkConfig

Networking options for a model explainability job.

modelExplainabilityBaselineConfig :: Maybe ModelExplainabilityBaselineConfig

The baseline configuration for a model explainability job.

tags :: Maybe [Tag]

(Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

stoppingCondition :: Maybe MonitoringStoppingCondition

jobDefinitionName :: Text

The name of the model explainability job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

modelExplainabilityAppSpecification :: ModelExplainabilityAppSpecification

Configures the model explainability job to run a specified Docker container image.

modelExplainabilityJobInput :: ModelExplainabilityJobInput

Inputs for the model explainability job.

modelExplainabilityJobOutputConfig :: MonitoringOutputConfig

jobResources :: MonitoringResources

roleArn :: Text

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

newCreateModelExplainabilityJobDefinition

Arguments

:: Text
$sel:jobDefinitionName:CreateModelExplainabilityJobDefinition'

-> ModelExplainabilityAppSpecification
$sel:modelExplainabilityAppSpecification:CreateModelExplainabilityJobDefinition'

-> ModelExplainabilityJobInput
$sel:modelExplainabilityJobInput:CreateModelExplainabilityJobDefinition'

-> MonitoringOutputConfig
$sel:modelExplainabilityJobOutputConfig:CreateModelExplainabilityJobDefinition'

-> MonitoringResources
$sel:jobResources:CreateModelExplainabilityJobDefinition'

-> Text
$sel:roleArn:CreateModelExplainabilityJobDefinition'

-> CreateModelExplainabilityJobDefinition

Create a value of CreateModelExplainabilityJobDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkConfig:CreateModelExplainabilityJobDefinition', createModelExplainabilityJobDefinition_networkConfig - Networking options for a model explainability job.

$sel:modelExplainabilityBaselineConfig:CreateModelExplainabilityJobDefinition', createModelExplainabilityJobDefinition_modelExplainabilityBaselineConfig - The baseline configuration for a model explainability job.

$sel:tags:CreateModelExplainabilityJobDefinition', createModelExplainabilityJobDefinition_tags - (Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

$sel:stoppingCondition:CreateModelExplainabilityJobDefinition', createModelExplainabilityJobDefinition_stoppingCondition - Undocumented member.

$sel:jobDefinitionName:CreateModelExplainabilityJobDefinition', createModelExplainabilityJobDefinition_jobDefinitionName - The name of the model explainability job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

$sel:modelExplainabilityAppSpecification:CreateModelExplainabilityJobDefinition', createModelExplainabilityJobDefinition_modelExplainabilityAppSpecification - Configures the model explainability job to run a specified Docker container image.

$sel:modelExplainabilityJobInput:CreateModelExplainabilityJobDefinition', createModelExplainabilityJobDefinition_modelExplainabilityJobInput - Inputs for the model explainability job.

$sel:modelExplainabilityJobOutputConfig:CreateModelExplainabilityJobDefinition', createModelExplainabilityJobDefinition_modelExplainabilityJobOutputConfig - Undocumented member.

$sel:jobResources:CreateModelExplainabilityJobDefinition', createModelExplainabilityJobDefinition_jobResources - Undocumented member.

$sel:roleArn:CreateModelExplainabilityJobDefinition', createModelExplainabilityJobDefinition_roleArn - The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

Request Lenses

createModelExplainabilityJobDefinition_networkConfig :: Lens' CreateModelExplainabilityJobDefinition (Maybe MonitoringNetworkConfig)

Networking options for a model explainability job.

createModelExplainabilityJobDefinition_modelExplainabilityBaselineConfig :: Lens' CreateModelExplainabilityJobDefinition (Maybe ModelExplainabilityBaselineConfig)

The baseline configuration for a model explainability job.

createModelExplainabilityJobDefinition_tags :: Lens' CreateModelExplainabilityJobDefinition (Maybe [Tag])

(Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

createModelExplainabilityJobDefinition_stoppingCondition :: Lens' CreateModelExplainabilityJobDefinition (Maybe MonitoringStoppingCondition)

Undocumented member.

createModelExplainabilityJobDefinition_jobDefinitionName :: Lens' CreateModelExplainabilityJobDefinition Text

The name of the model explainability job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

createModelExplainabilityJobDefinition_modelExplainabilityAppSpecification :: Lens' CreateModelExplainabilityJobDefinition ModelExplainabilityAppSpecification

Configures the model explainability job to run a specified Docker container image.

createModelExplainabilityJobDefinition_modelExplainabilityJobInput :: Lens' CreateModelExplainabilityJobDefinition ModelExplainabilityJobInput

Inputs for the model explainability job.

createModelExplainabilityJobDefinition_modelExplainabilityJobOutputConfig :: Lens' CreateModelExplainabilityJobDefinition MonitoringOutputConfig

Undocumented member.

createModelExplainabilityJobDefinition_jobResources :: Lens' CreateModelExplainabilityJobDefinition MonitoringResources

Undocumented member.

createModelExplainabilityJobDefinition_roleArn :: Lens' CreateModelExplainabilityJobDefinition Text

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

Destructuring the Response

data CreateModelExplainabilityJobDefinitionResponse

See: newCreateModelExplainabilityJobDefinitionResponse smart constructor.

Constructors

CreateModelExplainabilityJobDefinitionResponse'

Fields

httpStatus :: Int

The response's http status code.

jobDefinitionArn :: Text

The Amazon Resource Name (ARN) of the model explainability job.

newCreateModelExplainabilityJobDefinitionResponse

Arguments

:: Int
$sel:httpStatus:CreateModelExplainabilityJobDefinitionResponse'

-> Text
$sel:jobDefinitionArn:CreateModelExplainabilityJobDefinitionResponse'

-> CreateModelExplainabilityJobDefinitionResponse

Create a value of CreateModelExplainabilityJobDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateModelExplainabilityJobDefinitionResponse', createModelExplainabilityJobDefinitionResponse_httpStatus - The response's http status code.

$sel:jobDefinitionArn:CreateModelExplainabilityJobDefinitionResponse', createModelExplainabilityJobDefinitionResponse_jobDefinitionArn - The Amazon Resource Name (ARN) of the model explainability job.

Response Lenses

createModelExplainabilityJobDefinitionResponse_httpStatus :: Lens' CreateModelExplainabilityJobDefinitionResponse Int

The response's http status code.

createModelExplainabilityJobDefinitionResponse_jobDefinitionArn :: Lens' CreateModelExplainabilityJobDefinitionResponse Text

The Amazon Resource Name (ARN) of the model explainability job.