Network.AWS.SageMaker.CreateModelPackage

Description

Creates a model package that you can use to create Amazon SageMaker models or list on Amazon Web Services Marketplace, or a versioned model that is part of a model group. Buyers can subscribe to model packages listed on Amazon Web Services Marketplace to create models in Amazon SageMaker.

To create a model package by specifying a Docker container that contains your inference code and the Amazon S3 location of your model artifacts, provide values for InferenceSpecification. To create a model from an algorithm resource that you created or subscribed to in Amazon Web Services Marketplace, provide a value for SourceAlgorithmSpecification.

There are two types of model packages:

Versioned - a model that is part of a model group in the model registry.

Unversioned - a model package that is not part of a model group.

Creating a Request

data CreateModelPackage

See: newCreateModelPackage smart constructor.

Constructors

CreateModelPackage'

Fields

metadataProperties :: Maybe MetadataProperties

modelApprovalStatus :: Maybe ModelApprovalStatus

Whether the model is approved for deployment.

This parameter is optional for versioned models, and does not apply to unversioned models.

For versioned models, the value of this parameter must be set to Approved to deploy the model.

sourceAlgorithmSpecification :: Maybe SourceAlgorithmSpecification

Details about the algorithm that was used to create the model package.

modelPackageName :: Maybe Text

The name of the model package. The name must have 1 to 63 characters. Valid characters are a-z, A-Z, 0-9, and - (hyphen).

This parameter is required for unversioned models. It is not applicable to versioned models.

clientToken :: Maybe Text

A unique token that guarantees that the call to this API is idempotent.

modelMetrics :: Maybe ModelMetrics

A structure that contains model metrics reports.

modelPackageDescription :: Maybe Text

A description of the model package.

validationSpecification :: Maybe ModelPackageValidationSpecification

Specifies configurations for one or more transform jobs that Amazon SageMaker runs to test the model package.

inferenceSpecification :: Maybe InferenceSpecification

Specifies details about inference jobs that can be run with models based on this model package, including the following:

The Amazon ECR paths of containers that contain the inference code and model artifacts.

The instance types that the model package supports for transform jobs and real-time endpoints used for inference.

The input and output content formats that the model package supports for inference.

certifyForMarketplace :: Maybe Bool

Whether to certify the model package for listing on Amazon Web Services Marketplace.

This parameter is optional for unversioned models, and does not apply to versioned models.

modelPackageGroupName :: Maybe Text

The name of the model group that this model version belongs to.

This parameter is required for versioned models, and does not apply to unversioned models.

tags :: Maybe [Tag]

A list of key value pairs associated with the model. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

newCreateModelPackage :: CreateModelPackage

Create a value of CreateModelPackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metadataProperties:CreateModelPackage', createModelPackage_metadataProperties - Undocumented member.

$sel:modelApprovalStatus:CreateModelPackage', createModelPackage_modelApprovalStatus - Whether the model is approved for deployment.

This parameter is optional for versioned models, and does not apply to unversioned models.

For versioned models, the value of this parameter must be set to Approved to deploy the model.

$sel:sourceAlgorithmSpecification:CreateModelPackage', createModelPackage_sourceAlgorithmSpecification - Details about the algorithm that was used to create the model package.

$sel:modelPackageName:CreateModelPackage', createModelPackage_modelPackageName - The name of the model package. The name must have 1 to 63 characters. Valid characters are a-z, A-Z, 0-9, and - (hyphen).

This parameter is required for unversioned models. It is not applicable to versioned models.

$sel:clientToken:CreateModelPackage', createModelPackage_clientToken - A unique token that guarantees that the call to this API is idempotent.

$sel:modelMetrics:CreateModelPackage', createModelPackage_modelMetrics - A structure that contains model metrics reports.

$sel:modelPackageDescription:CreateModelPackage', createModelPackage_modelPackageDescription - A description of the model package.

$sel:validationSpecification:CreateModelPackage', createModelPackage_validationSpecification - Specifies configurations for one or more transform jobs that Amazon SageMaker runs to test the model package.

$sel:inferenceSpecification:CreateModelPackage', createModelPackage_inferenceSpecification - Specifies details about inference jobs that can be run with models based on this model package, including the following:

The Amazon ECR paths of containers that contain the inference code and model artifacts.

The instance types that the model package supports for transform jobs and real-time endpoints used for inference.

The input and output content formats that the model package supports for inference.

$sel:certifyForMarketplace:CreateModelPackage', createModelPackage_certifyForMarketplace - Whether to certify the model package for listing on Amazon Web Services Marketplace.

This parameter is optional for unversioned models, and does not apply to versioned models.

$sel:modelPackageGroupName:CreateModelPackage', createModelPackage_modelPackageGroupName - The name of the model group that this model version belongs to.

This parameter is required for versioned models, and does not apply to unversioned models.

$sel:tags:CreateModelPackage', createModelPackage_tags - A list of key value pairs associated with the model. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

Request Lenses

createModelPackage_metadataProperties :: Lens' CreateModelPackage (Maybe MetadataProperties)

Undocumented member.

createModelPackage_modelApprovalStatus :: Lens' CreateModelPackage (Maybe ModelApprovalStatus)

Whether the model is approved for deployment.

This parameter is optional for versioned models, and does not apply to unversioned models.

For versioned models, the value of this parameter must be set to Approved to deploy the model.

createModelPackage_sourceAlgorithmSpecification :: Lens' CreateModelPackage (Maybe SourceAlgorithmSpecification)

Details about the algorithm that was used to create the model package.

createModelPackage_modelPackageName :: Lens' CreateModelPackage (Maybe Text)

The name of the model package. The name must have 1 to 63 characters. Valid characters are a-z, A-Z, 0-9, and - (hyphen).

This parameter is required for unversioned models. It is not applicable to versioned models.

createModelPackage_clientToken :: Lens' CreateModelPackage (Maybe Text)

A unique token that guarantees that the call to this API is idempotent.

createModelPackage_modelMetrics :: Lens' CreateModelPackage (Maybe ModelMetrics)

A structure that contains model metrics reports.

createModelPackage_modelPackageDescription :: Lens' CreateModelPackage (Maybe Text)

A description of the model package.

createModelPackage_validationSpecification :: Lens' CreateModelPackage (Maybe ModelPackageValidationSpecification)

Specifies configurations for one or more transform jobs that Amazon SageMaker runs to test the model package.

createModelPackage_inferenceSpecification :: Lens' CreateModelPackage (Maybe InferenceSpecification)

Specifies details about inference jobs that can be run with models based on this model package, including the following:

The Amazon ECR paths of containers that contain the inference code and model artifacts.

The instance types that the model package supports for transform jobs and real-time endpoints used for inference.

The input and output content formats that the model package supports for inference.

createModelPackage_certifyForMarketplace :: Lens' CreateModelPackage (Maybe Bool)

Whether to certify the model package for listing on Amazon Web Services Marketplace.

This parameter is optional for unversioned models, and does not apply to versioned models.

createModelPackage_modelPackageGroupName :: Lens' CreateModelPackage (Maybe Text)

The name of the model group that this model version belongs to.

This parameter is required for versioned models, and does not apply to unversioned models.

createModelPackage_tags :: Lens' CreateModelPackage (Maybe [Tag])

A list of key value pairs associated with the model. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

Destructuring the Response

data CreateModelPackageResponse

See: newCreateModelPackageResponse smart constructor.

Constructors

CreateModelPackageResponse'

Fields

httpStatus :: Int

The response's http status code.

modelPackageArn :: Text

The Amazon Resource Name (ARN) of the new model package.

newCreateModelPackageResponse

Arguments

:: Int
$sel:httpStatus:CreateModelPackageResponse'

-> Text
$sel:modelPackageArn:CreateModelPackageResponse'

-> CreateModelPackageResponse

Create a value of CreateModelPackageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateModelPackageResponse', createModelPackageResponse_httpStatus - The response's http status code.

$sel:modelPackageArn:CreateModelPackageResponse', createModelPackageResponse_modelPackageArn - The Amazon Resource Name (ARN) of the new model package.

Response Lenses

createModelPackageResponse_httpStatus :: Lens' CreateModelPackageResponse Int

The response's http status code.

createModelPackageResponse_modelPackageArn :: Lens' CreateModelPackageResponse Text

The Amazon Resource Name (ARN) of the new model package.