Network.AWS.SageMaker.CreateNotebookInstance

Description

Creates an Amazon SageMaker notebook instance. A notebook instance is a machine learning (ML) compute instance running on a Jupyter notebook.

In a CreateNotebookInstance request, specify the type of ML compute instance that you want to run. Amazon SageMaker launches the instance, installs common libraries that you can use to explore datasets for model training, and attaches an ML storage volume to the notebook instance.

Amazon SageMaker also provides a set of example notebooks. Each notebook demonstrates how to use Amazon SageMaker with a specific algorithm or with a machine learning framework.

After receiving the request, Amazon SageMaker does the following:

Creates a network interface in the Amazon SageMaker VPC.

(Option) If you specified SubnetId, Amazon SageMaker creates a network interface in your own VPC, which is inferred from the subnet ID that you provide in the input. When creating this network interface, Amazon SageMaker attaches the security group that you specified in the request to the network interface that it creates in your VPC.

Launches an EC2 instance of the type specified in the request in the Amazon SageMaker VPC. If you specified SubnetId of your VPC, Amazon SageMaker specifies both network interfaces when launching this instance. This enables inbound traffic from your own VPC to the notebook instance, assuming that the security groups allow it.

After creating the notebook instance, Amazon SageMaker returns its Amazon Resource Name (ARN). You can't change the name of a notebook instance after you create it.

After Amazon SageMaker creates the notebook instance, you can connect to the Jupyter server and work in Jupyter notebooks. For example, you can write code to explore a dataset that you can use for model training, train a model, host models by creating Amazon SageMaker endpoints, and validate hosted models.

For more information, see How It Works.

Creating a Request

data CreateNotebookInstance

See: newCreateNotebookInstance smart constructor.

Constructors

CreateNotebookInstance'

Fields

securityGroupIds :: Maybe [Text]

The VPC security group IDs, in the form sg-xxxxxxxx. The security groups must be for the same VPC as specified in the subnet.

platformIdentifier :: Maybe Text

The platform identifier of the notebook instance runtime environment.

defaultCodeRepository :: Maybe Text

A Git repository to associate with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in Amazon Web Services CodeCommit or in any other Git repository. When you open a notebook instance, it opens in the directory that contains this repository. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

acceleratorTypes :: Maybe [NotebookInstanceAcceleratorType]

A list of Elastic Inference (EI) instance types to associate with this notebook instance. Currently, only one instance type can be associated with a notebook instance. For more information, see Using Elastic Inference in Amazon SageMaker.

additionalCodeRepositories :: Maybe [Text]

An array of up to three Git repositories to associate with the notebook instance. These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in Amazon Web Services CodeCommit or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

kmsKeyId :: Maybe Text

The Amazon Resource Name (ARN) of a Amazon Web Services Key Management Service key that Amazon SageMaker uses to encrypt data on the storage volume attached to your notebook instance. The KMS key you provide must be enabled. For information, see Enabling and Disabling Keys in the Amazon Web Services Key Management Service Developer Guide.

volumeSizeInGB :: Maybe Natural

The size, in GB, of the ML storage volume to attach to the notebook instance. The default value is 5 GB.

tags :: Maybe [Tag]

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

subnetId :: Maybe Text

The ID of the subnet in a VPC to which you would like to have a connectivity from your ML compute instance.

lifecycleConfigName :: Maybe Text

The name of a lifecycle configuration to associate with the notebook instance. For information about lifestyle configurations, see Step 2.1: (Optional) Customize a Notebook Instance.

rootAccess :: Maybe RootAccess

Whether root access is enabled or disabled for users of the notebook instance. The default value is Enabled.

Lifecycle configurations need root access to be able to set up a notebook instance. Because of this, lifecycle configurations associated with a notebook instance always run with root access even if you disable root access for users.

directInternetAccess :: Maybe DirectInternetAccess

Sets whether Amazon SageMaker provides internet access to the notebook instance. If you set this to Disabled this notebook instance is able to access resources only in your VPC, and is not be able to connect to Amazon SageMaker training and endpoint services unless you configure a NAT Gateway in your VPC.

For more information, see Notebook Instances Are Internet-Enabled by Default. You can set the value of this parameter to Disabled only if you set a value for the SubnetId parameter.

notebookInstanceName :: Text

The name of the new notebook instance.

instanceType :: InstanceType

The type of ML compute instance to launch for the notebook instance.

roleArn :: Text

When you send any requests to Amazon Web Services resources from the notebook instance, Amazon SageMaker assumes this role to perform tasks on your behalf. You must grant this role necessary permissions so Amazon SageMaker can perform these tasks. The policy must allow the Amazon SageMaker service principal (sagemaker.amazonaws.com) permissions to assume this role. For more information, see Amazon SageMaker Roles.

To be able to pass this role to Amazon SageMaker, the caller of this API must have the iam:PassRole permission.

newCreateNotebookInstance

Arguments

:: Text
$sel:notebookInstanceName:CreateNotebookInstance'

-> InstanceType
$sel:instanceType:CreateNotebookInstance'

-> Text
$sel:roleArn:CreateNotebookInstance'

-> CreateNotebookInstance

Create a value of CreateNotebookInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:CreateNotebookInstance', createNotebookInstance_securityGroupIds - The VPC security group IDs, in the form sg-xxxxxxxx. The security groups must be for the same VPC as specified in the subnet.

$sel:platformIdentifier:CreateNotebookInstance', createNotebookInstance_platformIdentifier - The platform identifier of the notebook instance runtime environment.

$sel:defaultCodeRepository:CreateNotebookInstance', createNotebookInstance_defaultCodeRepository - A Git repository to associate with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in Amazon Web Services CodeCommit or in any other Git repository. When you open a notebook instance, it opens in the directory that contains this repository. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

$sel:acceleratorTypes:CreateNotebookInstance', createNotebookInstance_acceleratorTypes - A list of Elastic Inference (EI) instance types to associate with this notebook instance. Currently, only one instance type can be associated with a notebook instance. For more information, see Using Elastic Inference in Amazon SageMaker.

$sel:additionalCodeRepositories:CreateNotebookInstance', createNotebookInstance_additionalCodeRepositories - An array of up to three Git repositories to associate with the notebook instance. These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in Amazon Web Services CodeCommit or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

$sel:kmsKeyId:CreateNotebookInstance', createNotebookInstance_kmsKeyId - The Amazon Resource Name (ARN) of a Amazon Web Services Key Management Service key that Amazon SageMaker uses to encrypt data on the storage volume attached to your notebook instance. The KMS key you provide must be enabled. For information, see Enabling and Disabling Keys in the Amazon Web Services Key Management Service Developer Guide.

$sel:volumeSizeInGB:CreateNotebookInstance', createNotebookInstance_volumeSizeInGB - The size, in GB, of the ML storage volume to attach to the notebook instance. The default value is 5 GB.

$sel:tags:CreateNotebookInstance', createNotebookInstance_tags - An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

$sel:subnetId:CreateNotebookInstance', createNotebookInstance_subnetId - The ID of the subnet in a VPC to which you would like to have a connectivity from your ML compute instance.

$sel:lifecycleConfigName:CreateNotebookInstance', createNotebookInstance_lifecycleConfigName - The name of a lifecycle configuration to associate with the notebook instance. For information about lifestyle configurations, see Step 2.1: (Optional) Customize a Notebook Instance.

$sel:rootAccess:CreateNotebookInstance', createNotebookInstance_rootAccess - Whether root access is enabled or disabled for users of the notebook instance. The default value is Enabled.

Lifecycle configurations need root access to be able to set up a notebook instance. Because of this, lifecycle configurations associated with a notebook instance always run with root access even if you disable root access for users.

$sel:directInternetAccess:CreateNotebookInstance', createNotebookInstance_directInternetAccess - Sets whether Amazon SageMaker provides internet access to the notebook instance. If you set this to Disabled this notebook instance is able to access resources only in your VPC, and is not be able to connect to Amazon SageMaker training and endpoint services unless you configure a NAT Gateway in your VPC.

For more information, see Notebook Instances Are Internet-Enabled by Default. You can set the value of this parameter to Disabled only if you set a value for the SubnetId parameter.

$sel:notebookInstanceName:CreateNotebookInstance', createNotebookInstance_notebookInstanceName - The name of the new notebook instance.

$sel:instanceType:CreateNotebookInstance', createNotebookInstance_instanceType - The type of ML compute instance to launch for the notebook instance.

$sel:roleArn:CreateNotebookInstance', createNotebookInstance_roleArn - When you send any requests to Amazon Web Services resources from the notebook instance, Amazon SageMaker assumes this role to perform tasks on your behalf. You must grant this role necessary permissions so Amazon SageMaker can perform these tasks. The policy must allow the Amazon SageMaker service principal (sagemaker.amazonaws.com) permissions to assume this role. For more information, see Amazon SageMaker Roles.

To be able to pass this role to Amazon SageMaker, the caller of this API must have the iam:PassRole permission.

Request Lenses

createNotebookInstance_securityGroupIds :: Lens' CreateNotebookInstance (Maybe [Text])

The VPC security group IDs, in the form sg-xxxxxxxx. The security groups must be for the same VPC as specified in the subnet.

createNotebookInstance_platformIdentifier :: Lens' CreateNotebookInstance (Maybe Text)

The platform identifier of the notebook instance runtime environment.

createNotebookInstance_defaultCodeRepository :: Lens' CreateNotebookInstance (Maybe Text)

A Git repository to associate with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in Amazon Web Services CodeCommit or in any other Git repository. When you open a notebook instance, it opens in the directory that contains this repository. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

createNotebookInstance_acceleratorTypes :: Lens' CreateNotebookInstance (Maybe [NotebookInstanceAcceleratorType])

A list of Elastic Inference (EI) instance types to associate with this notebook instance. Currently, only one instance type can be associated with a notebook instance. For more information, see Using Elastic Inference in Amazon SageMaker.

createNotebookInstance_additionalCodeRepositories :: Lens' CreateNotebookInstance (Maybe [Text])

An array of up to three Git repositories to associate with the notebook instance. These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in Amazon Web Services CodeCommit or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

createNotebookInstance_kmsKeyId :: Lens' CreateNotebookInstance (Maybe Text)

The Amazon Resource Name (ARN) of a Amazon Web Services Key Management Service key that Amazon SageMaker uses to encrypt data on the storage volume attached to your notebook instance. The KMS key you provide must be enabled. For information, see Enabling and Disabling Keys in the Amazon Web Services Key Management Service Developer Guide.

createNotebookInstance_volumeSizeInGB :: Lens' CreateNotebookInstance (Maybe Natural)

The size, in GB, of the ML storage volume to attach to the notebook instance. The default value is 5 GB.

createNotebookInstance_tags :: Lens' CreateNotebookInstance (Maybe [Tag])

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

createNotebookInstance_subnetId :: Lens' CreateNotebookInstance (Maybe Text)

The ID of the subnet in a VPC to which you would like to have a connectivity from your ML compute instance.

createNotebookInstance_lifecycleConfigName :: Lens' CreateNotebookInstance (Maybe Text)

The name of a lifecycle configuration to associate with the notebook instance. For information about lifestyle configurations, see Step 2.1: (Optional) Customize a Notebook Instance.

createNotebookInstance_rootAccess :: Lens' CreateNotebookInstance (Maybe RootAccess)

Whether root access is enabled or disabled for users of the notebook instance. The default value is Enabled.

Lifecycle configurations need root access to be able to set up a notebook instance. Because of this, lifecycle configurations associated with a notebook instance always run with root access even if you disable root access for users.

createNotebookInstance_directInternetAccess :: Lens' CreateNotebookInstance (Maybe DirectInternetAccess)

Sets whether Amazon SageMaker provides internet access to the notebook instance. If you set this to Disabled this notebook instance is able to access resources only in your VPC, and is not be able to connect to Amazon SageMaker training and endpoint services unless you configure a NAT Gateway in your VPC.

For more information, see Notebook Instances Are Internet-Enabled by Default. You can set the value of this parameter to Disabled only if you set a value for the SubnetId parameter.

createNotebookInstance_notebookInstanceName :: Lens' CreateNotebookInstance Text

The name of the new notebook instance.

createNotebookInstance_instanceType :: Lens' CreateNotebookInstance InstanceType

The type of ML compute instance to launch for the notebook instance.

createNotebookInstance_roleArn :: Lens' CreateNotebookInstance Text

When you send any requests to Amazon Web Services resources from the notebook instance, Amazon SageMaker assumes this role to perform tasks on your behalf. You must grant this role necessary permissions so Amazon SageMaker can perform these tasks. The policy must allow the Amazon SageMaker service principal (sagemaker.amazonaws.com) permissions to assume this role. For more information, see Amazon SageMaker Roles.

To be able to pass this role to Amazon SageMaker, the caller of this API must have the iam:PassRole permission.

Destructuring the Response

data CreateNotebookInstanceResponse

See: newCreateNotebookInstanceResponse smart constructor.

Constructors

CreateNotebookInstanceResponse'

Fields

notebookInstanceArn :: Maybe Text

The Amazon Resource Name (ARN) of the notebook instance.

httpStatus :: Int

The response's http status code.

newCreateNotebookInstanceResponse

Arguments

:: Int
$sel:httpStatus:CreateNotebookInstanceResponse'

-> CreateNotebookInstanceResponse

Create a value of CreateNotebookInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notebookInstanceArn:CreateNotebookInstanceResponse', createNotebookInstanceResponse_notebookInstanceArn - The Amazon Resource Name (ARN) of the notebook instance.

$sel:httpStatus:CreateNotebookInstanceResponse', createNotebookInstanceResponse_httpStatus - The response's http status code.

Response Lenses

createNotebookInstanceResponse_notebookInstanceArn :: Lens' CreateNotebookInstanceResponse (Maybe Text)

The Amazon Resource Name (ARN) of the notebook instance.

createNotebookInstanceResponse_httpStatus :: Lens' CreateNotebookInstanceResponse Int

The response's http status code.