Network.AWS.SageMaker.CreatePresignedDomainUrl

Description

Creates a URL for a specified UserProfile in a Domain. When accessed in a web browser, the user will be automatically signed in to Amazon SageMaker Studio, and granted access to all of the Apps and files associated with the Domain's Amazon Elastic File System (EFS) volume. This operation can only be called when the authentication mode equals IAM.

The IAM role or user used to call this API defines the permissions to access the app. Once the presigned URL is created, no additional permission is required to access this URL. IAM authorization policies for this API are also enforced for every HTTP request and WebSocket frame that attempts to connect to the app.

You can restrict access to this API and to the URL that it returns to a list of IP addresses, Amazon VPCs or Amazon VPC Endpoints that you specify. For more information, see Connect to SageMaker Studio Through an Interface VPC Endpoint .

The URL that you get from a call to CreatePresignedDomainUrl has a default timeout of 5 minutes. You can configure this value using ExpiresInSeconds. If you try to use the URL after the timeout limit expires, you are directed to the Amazon Web Services console sign-in page.

Creating a Request

data CreatePresignedDomainUrl

See: newCreatePresignedDomainUrl smart constructor.

Constructors

CreatePresignedDomainUrl'

Fields

sessionExpirationDurationInSeconds :: Maybe Natural

The session expiration duration in seconds. This value defaults to 43200.

expiresInSeconds :: Maybe Natural

The number of seconds until the pre-signed URL expires. This value defaults to 300.

domainId :: Text

The domain ID.

userProfileName :: Text

The name of the UserProfile to sign-in as.

newCreatePresignedDomainUrl

Arguments

:: Text
$sel:domainId:CreatePresignedDomainUrl'

-> Text
$sel:userProfileName:CreatePresignedDomainUrl'

-> CreatePresignedDomainUrl

Create a value of CreatePresignedDomainUrl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sessionExpirationDurationInSeconds:CreatePresignedDomainUrl', createPresignedDomainUrl_sessionExpirationDurationInSeconds - The session expiration duration in seconds. This value defaults to 43200.

$sel:expiresInSeconds:CreatePresignedDomainUrl', createPresignedDomainUrl_expiresInSeconds - The number of seconds until the pre-signed URL expires. This value defaults to 300.

$sel:domainId:CreatePresignedDomainUrl', createPresignedDomainUrl_domainId - The domain ID.

$sel:userProfileName:CreatePresignedDomainUrl', createPresignedDomainUrl_userProfileName - The name of the UserProfile to sign-in as.

Request Lenses

createPresignedDomainUrl_sessionExpirationDurationInSeconds :: Lens' CreatePresignedDomainUrl (Maybe Natural)

The session expiration duration in seconds. This value defaults to 43200.

createPresignedDomainUrl_expiresInSeconds :: Lens' CreatePresignedDomainUrl (Maybe Natural)

The number of seconds until the pre-signed URL expires. This value defaults to 300.

createPresignedDomainUrl_domainId :: Lens' CreatePresignedDomainUrl Text

The domain ID.

createPresignedDomainUrl_userProfileName :: Lens' CreatePresignedDomainUrl Text

The name of the UserProfile to sign-in as.

Destructuring the Response

data CreatePresignedDomainUrlResponse

See: newCreatePresignedDomainUrlResponse smart constructor.

Constructors

CreatePresignedDomainUrlResponse'

Fields

authorizedUrl :: Maybe Text

The presigned URL.

httpStatus :: Int

The response's http status code.

newCreatePresignedDomainUrlResponse

Arguments

:: Int
$sel:httpStatus:CreatePresignedDomainUrlResponse'

-> CreatePresignedDomainUrlResponse

Create a value of CreatePresignedDomainUrlResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorizedUrl:CreatePresignedDomainUrlResponse', createPresignedDomainUrlResponse_authorizedUrl - The presigned URL.

$sel:httpStatus:CreatePresignedDomainUrlResponse', createPresignedDomainUrlResponse_httpStatus - The response's http status code.

Response Lenses

createPresignedDomainUrlResponse_authorizedUrl :: Lens' CreatePresignedDomainUrlResponse (Maybe Text)

The presigned URL.

createPresignedDomainUrlResponse_httpStatus :: Lens' CreatePresignedDomainUrlResponse Int

The response's http status code.