Network.AWS.SageMaker.CreatePresignedNotebookInstanceUrl

Description

Returns a URL that you can use to connect to the Jupyter server from a notebook instance. In the Amazon SageMaker console, when you choose Open next to a notebook instance, Amazon SageMaker opens a new tab showing the Jupyter server home page from the notebook instance. The console uses this API to get the URL and show the page.

The IAM role or user used to call this API defines the permissions to access the notebook instance. Once the presigned URL is created, no additional permission is required to access this URL. IAM authorization policies for this API are also enforced for every HTTP request and WebSocket frame that attempts to connect to the notebook instance.

You can restrict access to this API and to the URL that it returns to a list of IP addresses that you specify. Use the NotIpAddress condition operator and the aws:SourceIP condition context key to specify the list of IP addresses that you want to have access to the notebook instance. For more information, see Limit Access to a Notebook Instance by IP Address.

The URL that you get from a call to CreatePresignedNotebookInstanceUrl is valid only for 5 minutes. If you try to use the URL after the 5-minute limit expires, you are directed to the Amazon Web Services console sign-in page.

Creating a Request

data CreatePresignedNotebookInstanceUrl

See: newCreatePresignedNotebookInstanceUrl smart constructor.

Constructors

CreatePresignedNotebookInstanceUrl'

Fields

sessionExpirationDurationInSeconds :: Maybe Natural

The duration of the session, in seconds. The default is 12 hours.

notebookInstanceName :: Text

The name of the notebook instance.

newCreatePresignedNotebookInstanceUrl

Arguments

:: Text
$sel:notebookInstanceName:CreatePresignedNotebookInstanceUrl'

-> CreatePresignedNotebookInstanceUrl

Create a value of CreatePresignedNotebookInstanceUrl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sessionExpirationDurationInSeconds:CreatePresignedNotebookInstanceUrl', createPresignedNotebookInstanceUrl_sessionExpirationDurationInSeconds - The duration of the session, in seconds. The default is 12 hours.

$sel:notebookInstanceName:CreatePresignedNotebookInstanceUrl', createPresignedNotebookInstanceUrl_notebookInstanceName - The name of the notebook instance.

Request Lenses

createPresignedNotebookInstanceUrl_sessionExpirationDurationInSeconds :: Lens' CreatePresignedNotebookInstanceUrl (Maybe Natural)

The duration of the session, in seconds. The default is 12 hours.

createPresignedNotebookInstanceUrl_notebookInstanceName :: Lens' CreatePresignedNotebookInstanceUrl Text

The name of the notebook instance.

Destructuring the Response

data CreatePresignedNotebookInstanceUrlResponse

See: newCreatePresignedNotebookInstanceUrlResponse smart constructor.

Constructors

CreatePresignedNotebookInstanceUrlResponse'

Fields

authorizedUrl :: Maybe Text

A JSON object that contains the URL string.

httpStatus :: Int

The response's http status code.

newCreatePresignedNotebookInstanceUrlResponse

Arguments

:: Int
$sel:httpStatus:CreatePresignedNotebookInstanceUrlResponse'

-> CreatePresignedNotebookInstanceUrlResponse

Create a value of CreatePresignedNotebookInstanceUrlResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorizedUrl:CreatePresignedNotebookInstanceUrlResponse', createPresignedNotebookInstanceUrlResponse_authorizedUrl - A JSON object that contains the URL string.

$sel:httpStatus:CreatePresignedNotebookInstanceUrlResponse', createPresignedNotebookInstanceUrlResponse_httpStatus - The response's http status code.

Response Lenses

createPresignedNotebookInstanceUrlResponse_authorizedUrl :: Lens' CreatePresignedNotebookInstanceUrlResponse (Maybe Text)

A JSON object that contains the URL string.

createPresignedNotebookInstanceUrlResponse_httpStatus :: Lens' CreatePresignedNotebookInstanceUrlResponse Int

The response's http status code.