Network.AWS.SageMaker.CreateProcessingJob

Description

Creates a processing job.

Creating a Request

data CreateProcessingJob

See: newCreateProcessingJob smart constructor.

Constructors

CreateProcessingJob'

Fields

environment :: Maybe (HashMap Text Text)

The environment variables to set in the Docker container. Up to 100 key and values entries in the map are supported.

stoppingCondition :: Maybe ProcessingStoppingCondition

The time limit for how long the processing job is allowed to run.

experimentConfig :: Maybe ExperimentConfig

processingInputs :: Maybe [ProcessingInput]

An array of inputs configuring the data to download into the processing container.

networkConfig :: Maybe NetworkConfig

Networking options for a processing job, such as whether to allow inbound and outbound network calls to and from processing containers, and the VPC subnets and security groups to use for VPC-enabled processing jobs.

processingOutputConfig :: Maybe ProcessingOutputConfig

Output configuration for the processing job.

tags :: Maybe [Tag]

(Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

processingJobName :: Text

The name of the processing job. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

processingResources :: ProcessingResources

Identifies the resources, ML compute instances, and ML storage volumes to deploy for a processing job. In distributed training, you specify more than one instance.

appSpecification :: AppSpecification

Configures the processing job to run a specified Docker container image.

roleArn :: Text

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

newCreateProcessingJob

Arguments

:: Text
$sel:processingJobName:CreateProcessingJob'

-> ProcessingResources
$sel:processingResources:CreateProcessingJob'

-> AppSpecification
$sel:appSpecification:CreateProcessingJob'

-> Text
$sel:roleArn:CreateProcessingJob'

-> CreateProcessingJob

Create a value of CreateProcessingJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environment:CreateProcessingJob', createProcessingJob_environment - The environment variables to set in the Docker container. Up to 100 key and values entries in the map are supported.

$sel:stoppingCondition:CreateProcessingJob', createProcessingJob_stoppingCondition - The time limit for how long the processing job is allowed to run.

$sel:experimentConfig:CreateProcessingJob', createProcessingJob_experimentConfig - Undocumented member.

$sel:processingInputs:CreateProcessingJob', createProcessingJob_processingInputs - An array of inputs configuring the data to download into the processing container.

$sel:networkConfig:CreateProcessingJob', createProcessingJob_networkConfig - Networking options for a processing job, such as whether to allow inbound and outbound network calls to and from processing containers, and the VPC subnets and security groups to use for VPC-enabled processing jobs.

$sel:processingOutputConfig:CreateProcessingJob', createProcessingJob_processingOutputConfig - Output configuration for the processing job.

$sel:tags:CreateProcessingJob', createProcessingJob_tags - (Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

$sel:processingJobName:CreateProcessingJob', createProcessingJob_processingJobName - The name of the processing job. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

$sel:processingResources:CreateProcessingJob', createProcessingJob_processingResources - Identifies the resources, ML compute instances, and ML storage volumes to deploy for a processing job. In distributed training, you specify more than one instance.

$sel:appSpecification:CreateProcessingJob', createProcessingJob_appSpecification - Configures the processing job to run a specified Docker container image.

$sel:roleArn:CreateProcessingJob', createProcessingJob_roleArn - The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

Request Lenses

createProcessingJob_environment :: Lens' CreateProcessingJob (Maybe (HashMap Text Text))

The environment variables to set in the Docker container. Up to 100 key and values entries in the map are supported.

createProcessingJob_stoppingCondition :: Lens' CreateProcessingJob (Maybe ProcessingStoppingCondition)

The time limit for how long the processing job is allowed to run.

createProcessingJob_experimentConfig :: Lens' CreateProcessingJob (Maybe ExperimentConfig)

Undocumented member.

createProcessingJob_processingInputs :: Lens' CreateProcessingJob (Maybe [ProcessingInput])

An array of inputs configuring the data to download into the processing container.

createProcessingJob_networkConfig :: Lens' CreateProcessingJob (Maybe NetworkConfig)

Networking options for a processing job, such as whether to allow inbound and outbound network calls to and from processing containers, and the VPC subnets and security groups to use for VPC-enabled processing jobs.

createProcessingJob_processingOutputConfig :: Lens' CreateProcessingJob (Maybe ProcessingOutputConfig)

Output configuration for the processing job.

createProcessingJob_tags :: Lens' CreateProcessingJob (Maybe [Tag])

(Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

createProcessingJob_processingJobName :: Lens' CreateProcessingJob Text

The name of the processing job. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

createProcessingJob_processingResources :: Lens' CreateProcessingJob ProcessingResources

Identifies the resources, ML compute instances, and ML storage volumes to deploy for a processing job. In distributed training, you specify more than one instance.

createProcessingJob_appSpecification :: Lens' CreateProcessingJob AppSpecification

Configures the processing job to run a specified Docker container image.

createProcessingJob_roleArn :: Lens' CreateProcessingJob Text

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

Destructuring the Response

data CreateProcessingJobResponse

See: newCreateProcessingJobResponse smart constructor.

Constructors

CreateProcessingJobResponse'

Fields

httpStatus :: Int

The response's http status code.

processingJobArn :: Text

The Amazon Resource Name (ARN) of the processing job.

newCreateProcessingJobResponse

Arguments

:: Int
$sel:httpStatus:CreateProcessingJobResponse'

-> Text
$sel:processingJobArn:CreateProcessingJobResponse'

-> CreateProcessingJobResponse

Create a value of CreateProcessingJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateProcessingJobResponse', createProcessingJobResponse_httpStatus - The response's http status code.

$sel:processingJobArn:CreateProcessingJobResponse', createProcessingJobResponse_processingJobArn - The Amazon Resource Name (ARN) of the processing job.

Response Lenses

createProcessingJobResponse_httpStatus :: Lens' CreateProcessingJobResponse Int

The response's http status code.

createProcessingJobResponse_processingJobArn :: Lens' CreateProcessingJobResponse Text

The Amazon Resource Name (ARN) of the processing job.