Network.AWS.SageMaker.CreateTrainingJob

Description

Starts a model training job. After training completes, Amazon SageMaker saves the resulting model artifacts to an Amazon S3 location that you specify.

If you choose to host your model using Amazon SageMaker hosting services, you can use the resulting model artifacts as part of the model. You can also use the artifacts in a machine learning service other than Amazon SageMaker, provided that you know how to use them for inference.

In the request body, you provide the following:

AlgorithmSpecification - Identifies the training algorithm to use.

HyperParameters - Specify these algorithm-specific parameters to enable the estimation of model parameters during training. Hyperparameters can be tuned to optimize this learning process. For a list of hyperparameters for each training algorithm provided by Amazon SageMaker, see Algorithms.

InputDataConfig - Describes the training dataset and the Amazon S3, EFS, or FSx location where it is stored.

OutputDataConfig - Identifies the Amazon S3 bucket where you want Amazon SageMaker to save the results of model training.

ResourceConfig - Identifies the resources, ML compute instances, and ML storage volumes to deploy for model training. In distributed training, you specify more than one instance.

EnableManagedSpotTraining - Optimize the cost of training machine learning models by up to 80% by using Amazon EC2 Spot instances. For more information, see Managed Spot Training.

RoleArn - The Amazon Resource Name (ARN) that Amazon SageMaker assumes to perform tasks on your behalf during model training. You must grant this role the necessary permissions so that Amazon SageMaker can successfully complete model training.

StoppingCondition - To help cap training costs, use MaxRuntimeInSeconds to set a time limit for training. Use MaxWaitTimeInSeconds to specify how long a managed spot training job has to complete.

Environment - The environment variables to set in the Docker container.

RetryStrategy - The number of times to retry the job when the job fails due to an InternalServerError.

For more information about Amazon SageMaker, see How It Works.

Creating a Request

data CreateTrainingJob

See: newCreateTrainingJob smart constructor.

Constructors

CreateTrainingJob'

Fields

environment :: Maybe (HashMap Text Text)

The environment variables to set in the Docker container.

debugHookConfig :: Maybe DebugHookConfig

checkpointConfig :: Maybe CheckpointConfig

Contains information about the output location for managed spot training checkpoint data.

retryStrategy :: Maybe RetryStrategy

The number of times to retry the job when the job fails due to an InternalServerError.

profilerConfig :: Maybe ProfilerConfig

enableNetworkIsolation :: Maybe Bool

Isolates the training container. No inbound or outbound network calls can be made, except for calls between peers within a training cluster for distributed training. If you enable network isolation for training jobs that are configured to use a VPC, Amazon SageMaker downloads and uploads customer data and model artifacts through the specified VPC, but the training container does not have network access.

experimentConfig :: Maybe ExperimentConfig

debugRuleConfigurations :: Maybe [DebugRuleConfiguration]

Configuration information for Debugger rules for debugging output tensors.

enableManagedSpotTraining :: Maybe Bool

To train models using managed spot training, choose True. Managed spot training provides a fully managed and scalable infrastructure for training machine learning models. this option is useful when training jobs can be interrupted and when there is flexibility when the training job is run.

The complete and intermediate results of jobs are stored in an Amazon S3 bucket, and can be used as a starting point to train models incrementally. Amazon SageMaker provides metrics and logs in CloudWatch. They can be used to see when managed spot training jobs are running, interrupted, resumed, or completed.

hyperParameters :: Maybe (HashMap Text Text)

Algorithm-specific parameters that influence the quality of the model. You set hyperparameters before you start the learning process. For a list of hyperparameters for each training algorithm provided by Amazon SageMaker, see Algorithms.

You can specify a maximum of 100 hyperparameters. Each hyperparameter is a key-value pair. Each key and value is limited to 256 characters, as specified by the Length Constraint.

inputDataConfig :: Maybe (NonEmpty Channel)

An array of Channel objects. Each channel is a named input source. InputDataConfig describes the input data and its location.

Algorithms can accept input data from one or more channels. For example, an algorithm might have two channels of input data, training_data and validation_data. The configuration for each channel provides the S3, EFS, or FSx location where the input data is stored. It also provides information about the stored data: the MIME type, compression method, and whether the data is wrapped in RecordIO format.

Depending on the input mode that the algorithm supports, Amazon SageMaker either copies input data files from an S3 bucket to a local directory in the Docker container, or makes it available as input streams. For example, if you specify an EFS location, input data files will be made available as input streams. They do not need to be downloaded.

profilerRuleConfigurations :: Maybe [ProfilerRuleConfiguration]

Configuration information for Debugger rules for profiling system and framework metrics.

vpcConfig :: Maybe VpcConfig

A VpcConfig object that specifies the VPC that you want your training job to connect to. Control access to and from your training container by configuring the VPC. For more information, see Protect Training Jobs by Using an Amazon Virtual Private Cloud.

enableInterContainerTrafficEncryption :: Maybe Bool

To encrypt all communications between ML compute instances in distributed training, choose True. Encryption provides greater security for distributed training, but training might take longer. How long it takes depends on the amount of communication between compute instances, especially if you use a deep learning algorithm in distributed training. For more information, see Protect Communications Between ML Compute Instances in a Distributed Training Job.

tensorBoardOutputConfig :: Maybe TensorBoardOutputConfig

tags :: Maybe [Tag]

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

trainingJobName :: Text

The name of the training job. The name must be unique within an Amazon Web Services Region in an Amazon Web Services account.

algorithmSpecification :: AlgorithmSpecification

The registry path of the Docker image that contains the training algorithm and algorithm-specific metadata, including the input mode. For more information about algorithms provided by Amazon SageMaker, see Algorithms. For information about providing your own algorithms, see Using Your Own Algorithms with Amazon SageMaker.

roleArn :: Text

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

During model training, Amazon SageMaker needs your permission to read input data from an S3 bucket, download a Docker image that contains training code, write model artifacts to an S3 bucket, write logs to Amazon CloudWatch Logs, and publish metrics to Amazon CloudWatch. You grant permissions for all of these tasks to an IAM role. For more information, see Amazon SageMaker Roles.

To be able to pass this role to Amazon SageMaker, the caller of this API must have the iam:PassRole permission.

outputDataConfig :: OutputDataConfig

Specifies the path to the S3 location where you want to store model artifacts. Amazon SageMaker creates subfolders for the artifacts.

resourceConfig :: ResourceConfig

The resources, including the ML compute instances and ML storage volumes, to use for model training.

ML storage volumes store model artifacts and incremental states. Training algorithms might also use ML storage volumes for scratch space. If you want Amazon SageMaker to use the ML storage volume to store the training data, choose File as the TrainingInputMode in the algorithm specification. For distributed training algorithms, specify an instance count greater than 1.

stoppingCondition :: StoppingCondition

Specifies a limit to how long a model training job can run. It also specifies how long a managed Spot training job has to complete. When the job reaches the time limit, Amazon SageMaker ends the training job. Use this API to cap model training costs.

To stop a job, Amazon SageMaker sends the algorithm the SIGTERM signal, which delays job termination for 120 seconds. Algorithms can use this 120-second window to save the model artifacts, so the results of training are not lost.

newCreateTrainingJob

Arguments

:: Text
$sel:trainingJobName:CreateTrainingJob'

-> AlgorithmSpecification
$sel:algorithmSpecification:CreateTrainingJob'

-> Text
$sel:roleArn:CreateTrainingJob'

-> OutputDataConfig
$sel:outputDataConfig:CreateTrainingJob'

-> ResourceConfig
$sel:resourceConfig:CreateTrainingJob'

-> StoppingCondition
$sel:stoppingCondition:CreateTrainingJob'

-> CreateTrainingJob

Create a value of CreateTrainingJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environment:CreateTrainingJob', createTrainingJob_environment - The environment variables to set in the Docker container.

$sel:debugHookConfig:CreateTrainingJob', createTrainingJob_debugHookConfig - Undocumented member.

$sel:checkpointConfig:CreateTrainingJob', createTrainingJob_checkpointConfig - Contains information about the output location for managed spot training checkpoint data.

$sel:retryStrategy:CreateTrainingJob', createTrainingJob_retryStrategy - The number of times to retry the job when the job fails due to an InternalServerError.

$sel:profilerConfig:CreateTrainingJob', createTrainingJob_profilerConfig - Undocumented member.

$sel:enableNetworkIsolation:CreateTrainingJob', createTrainingJob_enableNetworkIsolation - Isolates the training container. No inbound or outbound network calls can be made, except for calls between peers within a training cluster for distributed training. If you enable network isolation for training jobs that are configured to use a VPC, Amazon SageMaker downloads and uploads customer data and model artifacts through the specified VPC, but the training container does not have network access.

$sel:experimentConfig:CreateTrainingJob', createTrainingJob_experimentConfig - Undocumented member.

$sel:debugRuleConfigurations:CreateTrainingJob', createTrainingJob_debugRuleConfigurations - Configuration information for Debugger rules for debugging output tensors.

$sel:enableManagedSpotTraining:CreateTrainingJob', createTrainingJob_enableManagedSpotTraining - To train models using managed spot training, choose True. Managed spot training provides a fully managed and scalable infrastructure for training machine learning models. this option is useful when training jobs can be interrupted and when there is flexibility when the training job is run.

The complete and intermediate results of jobs are stored in an Amazon S3 bucket, and can be used as a starting point to train models incrementally. Amazon SageMaker provides metrics and logs in CloudWatch. They can be used to see when managed spot training jobs are running, interrupted, resumed, or completed.

$sel:hyperParameters:CreateTrainingJob', createTrainingJob_hyperParameters - Algorithm-specific parameters that influence the quality of the model. You set hyperparameters before you start the learning process. For a list of hyperparameters for each training algorithm provided by Amazon SageMaker, see Algorithms.

You can specify a maximum of 100 hyperparameters. Each hyperparameter is a key-value pair. Each key and value is limited to 256 characters, as specified by the Length Constraint.

$sel:inputDataConfig:CreateTrainingJob', createTrainingJob_inputDataConfig - An array of Channel objects. Each channel is a named input source. InputDataConfig describes the input data and its location.

Algorithms can accept input data from one or more channels. For example, an algorithm might have two channels of input data, training_data and validation_data. The configuration for each channel provides the S3, EFS, or FSx location where the input data is stored. It also provides information about the stored data: the MIME type, compression method, and whether the data is wrapped in RecordIO format.

Depending on the input mode that the algorithm supports, Amazon SageMaker either copies input data files from an S3 bucket to a local directory in the Docker container, or makes it available as input streams. For example, if you specify an EFS location, input data files will be made available as input streams. They do not need to be downloaded.

$sel:profilerRuleConfigurations:CreateTrainingJob', createTrainingJob_profilerRuleConfigurations - Configuration information for Debugger rules for profiling system and framework metrics.

$sel:vpcConfig:CreateTrainingJob', createTrainingJob_vpcConfig - A VpcConfig object that specifies the VPC that you want your training job to connect to. Control access to and from your training container by configuring the VPC. For more information, see Protect Training Jobs by Using an Amazon Virtual Private Cloud.

$sel:enableInterContainerTrafficEncryption:CreateTrainingJob', createTrainingJob_enableInterContainerTrafficEncryption - To encrypt all communications between ML compute instances in distributed training, choose True. Encryption provides greater security for distributed training, but training might take longer. How long it takes depends on the amount of communication between compute instances, especially if you use a deep learning algorithm in distributed training. For more information, see Protect Communications Between ML Compute Instances in a Distributed Training Job.

$sel:tensorBoardOutputConfig:CreateTrainingJob', createTrainingJob_tensorBoardOutputConfig - Undocumented member.

$sel:tags:CreateTrainingJob', createTrainingJob_tags - An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

$sel:trainingJobName:CreateTrainingJob', createTrainingJob_trainingJobName - The name of the training job. The name must be unique within an Amazon Web Services Region in an Amazon Web Services account.

$sel:algorithmSpecification:CreateTrainingJob', createTrainingJob_algorithmSpecification - The registry path of the Docker image that contains the training algorithm and algorithm-specific metadata, including the input mode. For more information about algorithms provided by Amazon SageMaker, see Algorithms. For information about providing your own algorithms, see Using Your Own Algorithms with Amazon SageMaker.

$sel:roleArn:CreateTrainingJob', createTrainingJob_roleArn - The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

During model training, Amazon SageMaker needs your permission to read input data from an S3 bucket, download a Docker image that contains training code, write model artifacts to an S3 bucket, write logs to Amazon CloudWatch Logs, and publish metrics to Amazon CloudWatch. You grant permissions for all of these tasks to an IAM role. For more information, see Amazon SageMaker Roles.

To be able to pass this role to Amazon SageMaker, the caller of this API must have the iam:PassRole permission.

$sel:outputDataConfig:CreateTrainingJob', createTrainingJob_outputDataConfig - Specifies the path to the S3 location where you want to store model artifacts. Amazon SageMaker creates subfolders for the artifacts.

$sel:resourceConfig:CreateTrainingJob', createTrainingJob_resourceConfig - The resources, including the ML compute instances and ML storage volumes, to use for model training.

ML storage volumes store model artifacts and incremental states. Training algorithms might also use ML storage volumes for scratch space. If you want Amazon SageMaker to use the ML storage volume to store the training data, choose File as the TrainingInputMode in the algorithm specification. For distributed training algorithms, specify an instance count greater than 1.

$sel:stoppingCondition:CreateTrainingJob', createTrainingJob_stoppingCondition - Specifies a limit to how long a model training job can run. It also specifies how long a managed Spot training job has to complete. When the job reaches the time limit, Amazon SageMaker ends the training job. Use this API to cap model training costs.

To stop a job, Amazon SageMaker sends the algorithm the SIGTERM signal, which delays job termination for 120 seconds. Algorithms can use this 120-second window to save the model artifacts, so the results of training are not lost.

Request Lenses

createTrainingJob_environment :: Lens' CreateTrainingJob (Maybe (HashMap Text Text))

The environment variables to set in the Docker container.

createTrainingJob_debugHookConfig :: Lens' CreateTrainingJob (Maybe DebugHookConfig)

Undocumented member.

createTrainingJob_checkpointConfig :: Lens' CreateTrainingJob (Maybe CheckpointConfig)

Contains information about the output location for managed spot training checkpoint data.

createTrainingJob_retryStrategy :: Lens' CreateTrainingJob (Maybe RetryStrategy)

The number of times to retry the job when the job fails due to an InternalServerError.

createTrainingJob_profilerConfig :: Lens' CreateTrainingJob (Maybe ProfilerConfig)

Undocumented member.

createTrainingJob_enableNetworkIsolation :: Lens' CreateTrainingJob (Maybe Bool)

Isolates the training container. No inbound or outbound network calls can be made, except for calls between peers within a training cluster for distributed training. If you enable network isolation for training jobs that are configured to use a VPC, Amazon SageMaker downloads and uploads customer data and model artifacts through the specified VPC, but the training container does not have network access.

createTrainingJob_experimentConfig :: Lens' CreateTrainingJob (Maybe ExperimentConfig)

Undocumented member.

createTrainingJob_debugRuleConfigurations :: Lens' CreateTrainingJob (Maybe [DebugRuleConfiguration])

Configuration information for Debugger rules for debugging output tensors.

createTrainingJob_enableManagedSpotTraining :: Lens' CreateTrainingJob (Maybe Bool)

To train models using managed spot training, choose True. Managed spot training provides a fully managed and scalable infrastructure for training machine learning models. this option is useful when training jobs can be interrupted and when there is flexibility when the training job is run.

The complete and intermediate results of jobs are stored in an Amazon S3 bucket, and can be used as a starting point to train models incrementally. Amazon SageMaker provides metrics and logs in CloudWatch. They can be used to see when managed spot training jobs are running, interrupted, resumed, or completed.

createTrainingJob_hyperParameters :: Lens' CreateTrainingJob (Maybe (HashMap Text Text))

Algorithm-specific parameters that influence the quality of the model. You set hyperparameters before you start the learning process. For a list of hyperparameters for each training algorithm provided by Amazon SageMaker, see Algorithms.

You can specify a maximum of 100 hyperparameters. Each hyperparameter is a key-value pair. Each key and value is limited to 256 characters, as specified by the Length Constraint.

createTrainingJob_inputDataConfig :: Lens' CreateTrainingJob (Maybe (NonEmpty Channel))

An array of Channel objects. Each channel is a named input source. InputDataConfig describes the input data and its location.

Algorithms can accept input data from one or more channels. For example, an algorithm might have two channels of input data, training_data and validation_data. The configuration for each channel provides the S3, EFS, or FSx location where the input data is stored. It also provides information about the stored data: the MIME type, compression method, and whether the data is wrapped in RecordIO format.

Depending on the input mode that the algorithm supports, Amazon SageMaker either copies input data files from an S3 bucket to a local directory in the Docker container, or makes it available as input streams. For example, if you specify an EFS location, input data files will be made available as input streams. They do not need to be downloaded.

createTrainingJob_profilerRuleConfigurations :: Lens' CreateTrainingJob (Maybe [ProfilerRuleConfiguration])

Configuration information for Debugger rules for profiling system and framework metrics.

createTrainingJob_vpcConfig :: Lens' CreateTrainingJob (Maybe VpcConfig)

A VpcConfig object that specifies the VPC that you want your training job to connect to. Control access to and from your training container by configuring the VPC. For more information, see Protect Training Jobs by Using an Amazon Virtual Private Cloud.

createTrainingJob_enableInterContainerTrafficEncryption :: Lens' CreateTrainingJob (Maybe Bool)

To encrypt all communications between ML compute instances in distributed training, choose True. Encryption provides greater security for distributed training, but training might take longer. How long it takes depends on the amount of communication between compute instances, especially if you use a deep learning algorithm in distributed training. For more information, see Protect Communications Between ML Compute Instances in a Distributed Training Job.

createTrainingJob_tensorBoardOutputConfig :: Lens' CreateTrainingJob (Maybe TensorBoardOutputConfig)

Undocumented member.

createTrainingJob_tags :: Lens' CreateTrainingJob (Maybe [Tag])

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

createTrainingJob_trainingJobName :: Lens' CreateTrainingJob Text

The name of the training job. The name must be unique within an Amazon Web Services Region in an Amazon Web Services account.

createTrainingJob_algorithmSpecification :: Lens' CreateTrainingJob AlgorithmSpecification

The registry path of the Docker image that contains the training algorithm and algorithm-specific metadata, including the input mode. For more information about algorithms provided by Amazon SageMaker, see Algorithms. For information about providing your own algorithms, see Using Your Own Algorithms with Amazon SageMaker.

createTrainingJob_roleArn :: Lens' CreateTrainingJob Text

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

During model training, Amazon SageMaker needs your permission to read input data from an S3 bucket, download a Docker image that contains training code, write model artifacts to an S3 bucket, write logs to Amazon CloudWatch Logs, and publish metrics to Amazon CloudWatch. You grant permissions for all of these tasks to an IAM role. For more information, see Amazon SageMaker Roles.

To be able to pass this role to Amazon SageMaker, the caller of this API must have the iam:PassRole permission.

createTrainingJob_outputDataConfig :: Lens' CreateTrainingJob OutputDataConfig

Specifies the path to the S3 location where you want to store model artifacts. Amazon SageMaker creates subfolders for the artifacts.

createTrainingJob_resourceConfig :: Lens' CreateTrainingJob ResourceConfig

The resources, including the ML compute instances and ML storage volumes, to use for model training.

ML storage volumes store model artifacts and incremental states. Training algorithms might also use ML storage volumes for scratch space. If you want Amazon SageMaker to use the ML storage volume to store the training data, choose File as the TrainingInputMode in the algorithm specification. For distributed training algorithms, specify an instance count greater than 1.

createTrainingJob_stoppingCondition :: Lens' CreateTrainingJob StoppingCondition

Specifies a limit to how long a model training job can run. It also specifies how long a managed Spot training job has to complete. When the job reaches the time limit, Amazon SageMaker ends the training job. Use this API to cap model training costs.

To stop a job, Amazon SageMaker sends the algorithm the SIGTERM signal, which delays job termination for 120 seconds. Algorithms can use this 120-second window to save the model artifacts, so the results of training are not lost.

Destructuring the Response

data CreateTrainingJobResponse

See: newCreateTrainingJobResponse smart constructor.

Constructors

CreateTrainingJobResponse'

Fields

httpStatus :: Int

The response's http status code.

trainingJobArn :: Text

The Amazon Resource Name (ARN) of the training job.

newCreateTrainingJobResponse

Arguments

:: Int
$sel:httpStatus:CreateTrainingJobResponse'

-> Text
$sel:trainingJobArn:CreateTrainingJobResponse'

-> CreateTrainingJobResponse

Create a value of CreateTrainingJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateTrainingJobResponse', createTrainingJobResponse_httpStatus - The response's http status code.

$sel:trainingJobArn:CreateTrainingJobResponse', createTrainingJobResponse_trainingJobArn - The Amazon Resource Name (ARN) of the training job.

Response Lenses

createTrainingJobResponse_httpStatus :: Lens' CreateTrainingJobResponse Int

The response's http status code.

createTrainingJobResponse_trainingJobArn :: Lens' CreateTrainingJobResponse Text

The Amazon Resource Name (ARN) of the training job.