Network.AWS.SageMaker.CreateTransformJob

Description

Starts a transform job. A transform job uses a trained model to get inferences on a dataset and saves these results to an Amazon S3 location that you specify.

To perform batch transformations, you create a transform job and use the data that you have readily available.

In the request body, you provide the following:

TransformJobName - Identifies the transform job. The name must be unique within an Amazon Web Services Region in an Amazon Web Services account.

ModelName - Identifies the model to use. ModelName must be the name of an existing Amazon SageMaker model in the same Amazon Web Services Region and Amazon Web Services account. For information on creating a model, see CreateModel.

TransformInput - Describes the dataset to be transformed and the Amazon S3 location where it is stored.

TransformOutput - Identifies the Amazon S3 location where you want Amazon SageMaker to save the results from the transform job.

TransformResources - Identifies the ML compute instances for the transform job.

For more information about how batch transformation works, see Batch Transform.

Creating a Request

data CreateTransformJob

See: newCreateTransformJob smart constructor.

Constructors

CreateTransformJob'

Fields

experimentConfig :: Maybe ExperimentConfig

maxConcurrentTransforms :: Maybe Natural

The maximum number of parallel requests that can be sent to each instance in a transform job. If MaxConcurrentTransforms is set to 0 or left unset, Amazon SageMaker checks the optional execution-parameters to determine the settings for your chosen algorithm. If the execution-parameters endpoint is not enabled, the default value is 1. For more information on execution-parameters, see How Containers Serve Requests. For built-in algorithms, you don't need to set a value for MaxConcurrentTransforms.

environment :: Maybe (HashMap Text Text)

The environment variables to set in the Docker container. We support up to 16 key and values entries in the map.

maxPayloadInMB :: Maybe Natural

The maximum allowed size of the payload, in MB. A payload is the data portion of a record (without metadata). The value in MaxPayloadInMB must be greater than, or equal to, the size of a single record. To estimate the size of a record in MB, divide the size of your dataset by the number of records. To ensure that the records fit within the maximum payload size, we recommend using a slightly larger value. The default value is 6 MB.

For cases where the payload might be arbitrarily large and is transmitted using HTTP chunked encoding, set the value to 0. This feature works only in supported algorithms. Currently, Amazon SageMaker built-in algorithms do not support HTTP chunked encoding.

batchStrategy :: Maybe BatchStrategy

Specifies the number of records to include in a mini-batch for an HTTP inference request. A record // is a single unit of input data that inference can be made on. For example, a single line in a CSV file is a record.

To enable the batch strategy, you must set the SplitType property to Line, RecordIO, or TFRecord.

To use only one record when making an HTTP invocation request to a container, set BatchStrategy to SingleRecord and SplitType to Line.

To fit as many records in a mini-batch as can fit within the MaxPayloadInMB limit, set BatchStrategy to MultiRecord and SplitType to Line.

modelClientConfig :: Maybe ModelClientConfig

Configures the timeout and maximum number of retries for processing a transform job invocation.

tags :: Maybe [Tag]

(Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

dataProcessing :: Maybe DataProcessing

The data structure used to specify the data to be used for inference in a batch transform job and to associate the data that is relevant to the prediction results in the output. The input filter provided allows you to exclude input data that is not needed for inference in a batch transform job. The output filter provided allows you to include input data relevant to interpreting the predictions in the output from the job. For more information, see Associate Prediction Results with their Corresponding Input Records.

transformJobName :: Text

The name of the transform job. The name must be unique within an Amazon Web Services Region in an Amazon Web Services account.

modelName :: Text

The name of the model that you want to use for the transform job. ModelName must be the name of an existing Amazon SageMaker model within an Amazon Web Services Region in an Amazon Web Services account.

transformInput :: TransformInput

Describes the input source and the way the transform job consumes it.

transformOutput :: TransformOutput

Describes the results of the transform job.

transformResources :: TransformResources

Describes the resources, including ML instance types and ML instance count, to use for the transform job.

newCreateTransformJob

Arguments

:: Text
$sel:transformJobName:CreateTransformJob'

-> Text
$sel:modelName:CreateTransformJob'

-> TransformInput
$sel:transformInput:CreateTransformJob'

-> TransformOutput
$sel:transformOutput:CreateTransformJob'

-> TransformResources
$sel:transformResources:CreateTransformJob'

-> CreateTransformJob

Create a value of CreateTransformJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:experimentConfig:CreateTransformJob', createTransformJob_experimentConfig - Undocumented member.

$sel:maxConcurrentTransforms:CreateTransformJob', createTransformJob_maxConcurrentTransforms - The maximum number of parallel requests that can be sent to each instance in a transform job. If MaxConcurrentTransforms is set to 0 or left unset, Amazon SageMaker checks the optional execution-parameters to determine the settings for your chosen algorithm. If the execution-parameters endpoint is not enabled, the default value is 1. For more information on execution-parameters, see How Containers Serve Requests. For built-in algorithms, you don't need to set a value for MaxConcurrentTransforms.

$sel:environment:CreateTransformJob', createTransformJob_environment - The environment variables to set in the Docker container. We support up to 16 key and values entries in the map.

$sel:maxPayloadInMB:CreateTransformJob', createTransformJob_maxPayloadInMB - The maximum allowed size of the payload, in MB. A payload is the data portion of a record (without metadata). The value in MaxPayloadInMB must be greater than, or equal to, the size of a single record. To estimate the size of a record in MB, divide the size of your dataset by the number of records. To ensure that the records fit within the maximum payload size, we recommend using a slightly larger value. The default value is 6 MB.

For cases where the payload might be arbitrarily large and is transmitted using HTTP chunked encoding, set the value to 0. This feature works only in supported algorithms. Currently, Amazon SageMaker built-in algorithms do not support HTTP chunked encoding.

$sel:batchStrategy:CreateTransformJob', createTransformJob_batchStrategy - Specifies the number of records to include in a mini-batch for an HTTP inference request. A record // is a single unit of input data that inference can be made on. For example, a single line in a CSV file is a record.

To enable the batch strategy, you must set the SplitType property to Line, RecordIO, or TFRecord.

To use only one record when making an HTTP invocation request to a container, set BatchStrategy to SingleRecord and SplitType to Line.

To fit as many records in a mini-batch as can fit within the MaxPayloadInMB limit, set BatchStrategy to MultiRecord and SplitType to Line.

$sel:modelClientConfig:CreateTransformJob', createTransformJob_modelClientConfig - Configures the timeout and maximum number of retries for processing a transform job invocation.

$sel:tags:CreateTransformJob', createTransformJob_tags - (Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

$sel:dataProcessing:CreateTransformJob', createTransformJob_dataProcessing - The data structure used to specify the data to be used for inference in a batch transform job and to associate the data that is relevant to the prediction results in the output. The input filter provided allows you to exclude input data that is not needed for inference in a batch transform job. The output filter provided allows you to include input data relevant to interpreting the predictions in the output from the job. For more information, see Associate Prediction Results with their Corresponding Input Records.

$sel:transformJobName:CreateTransformJob', createTransformJob_transformJobName - The name of the transform job. The name must be unique within an Amazon Web Services Region in an Amazon Web Services account.

$sel:modelName:CreateTransformJob', createTransformJob_modelName - The name of the model that you want to use for the transform job. ModelName must be the name of an existing Amazon SageMaker model within an Amazon Web Services Region in an Amazon Web Services account.

$sel:transformInput:CreateTransformJob', createTransformJob_transformInput - Describes the input source and the way the transform job consumes it.

$sel:transformOutput:CreateTransformJob', createTransformJob_transformOutput - Describes the results of the transform job.

$sel:transformResources:CreateTransformJob', createTransformJob_transformResources - Describes the resources, including ML instance types and ML instance count, to use for the transform job.

Request Lenses

createTransformJob_experimentConfig :: Lens' CreateTransformJob (Maybe ExperimentConfig)

Undocumented member.

createTransformJob_maxConcurrentTransforms :: Lens' CreateTransformJob (Maybe Natural)

The maximum number of parallel requests that can be sent to each instance in a transform job. If MaxConcurrentTransforms is set to 0 or left unset, Amazon SageMaker checks the optional execution-parameters to determine the settings for your chosen algorithm. If the execution-parameters endpoint is not enabled, the default value is 1. For more information on execution-parameters, see How Containers Serve Requests. For built-in algorithms, you don't need to set a value for MaxConcurrentTransforms.

createTransformJob_environment :: Lens' CreateTransformJob (Maybe (HashMap Text Text))

The environment variables to set in the Docker container. We support up to 16 key and values entries in the map.

createTransformJob_maxPayloadInMB :: Lens' CreateTransformJob (Maybe Natural)

The maximum allowed size of the payload, in MB. A payload is the data portion of a record (without metadata). The value in MaxPayloadInMB must be greater than, or equal to, the size of a single record. To estimate the size of a record in MB, divide the size of your dataset by the number of records. To ensure that the records fit within the maximum payload size, we recommend using a slightly larger value. The default value is 6 MB.

For cases where the payload might be arbitrarily large and is transmitted using HTTP chunked encoding, set the value to 0. This feature works only in supported algorithms. Currently, Amazon SageMaker built-in algorithms do not support HTTP chunked encoding.

createTransformJob_batchStrategy :: Lens' CreateTransformJob (Maybe BatchStrategy)

Specifies the number of records to include in a mini-batch for an HTTP inference request. A record // is a single unit of input data that inference can be made on. For example, a single line in a CSV file is a record.

To enable the batch strategy, you must set the SplitType property to Line, RecordIO, or TFRecord.

To use only one record when making an HTTP invocation request to a container, set BatchStrategy to SingleRecord and SplitType to Line.

To fit as many records in a mini-batch as can fit within the MaxPayloadInMB limit, set BatchStrategy to MultiRecord and SplitType to Line.

createTransformJob_modelClientConfig :: Lens' CreateTransformJob (Maybe ModelClientConfig)

Configures the timeout and maximum number of retries for processing a transform job invocation.

createTransformJob_tags :: Lens' CreateTransformJob (Maybe [Tag])

(Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

createTransformJob_dataProcessing :: Lens' CreateTransformJob (Maybe DataProcessing)

The data structure used to specify the data to be used for inference in a batch transform job and to associate the data that is relevant to the prediction results in the output. The input filter provided allows you to exclude input data that is not needed for inference in a batch transform job. The output filter provided allows you to include input data relevant to interpreting the predictions in the output from the job. For more information, see Associate Prediction Results with their Corresponding Input Records.

createTransformJob_transformJobName :: Lens' CreateTransformJob Text

The name of the transform job. The name must be unique within an Amazon Web Services Region in an Amazon Web Services account.

createTransformJob_modelName :: Lens' CreateTransformJob Text

The name of the model that you want to use for the transform job. ModelName must be the name of an existing Amazon SageMaker model within an Amazon Web Services Region in an Amazon Web Services account.

createTransformJob_transformInput :: Lens' CreateTransformJob TransformInput

Describes the input source and the way the transform job consumes it.

createTransformJob_transformOutput :: Lens' CreateTransformJob TransformOutput

Describes the results of the transform job.

createTransformJob_transformResources :: Lens' CreateTransformJob TransformResources

Describes the resources, including ML instance types and ML instance count, to use for the transform job.

Destructuring the Response

data CreateTransformJobResponse

See: newCreateTransformJobResponse smart constructor.

Constructors

CreateTransformJobResponse'

Fields

httpStatus :: Int

The response's http status code.

transformJobArn :: Text

The Amazon Resource Name (ARN) of the transform job.

newCreateTransformJobResponse

Arguments

:: Int
$sel:httpStatus:CreateTransformJobResponse'

-> Text
$sel:transformJobArn:CreateTransformJobResponse'

-> CreateTransformJobResponse

Create a value of CreateTransformJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateTransformJobResponse', createTransformJobResponse_httpStatus - The response's http status code.

$sel:transformJobArn:CreateTransformJobResponse', createTransformJobResponse_transformJobArn - The Amazon Resource Name (ARN) of the transform job.

Response Lenses

createTransformJobResponse_httpStatus :: Lens' CreateTransformJobResponse Int

The response's http status code.

createTransformJobResponse_transformJobArn :: Lens' CreateTransformJobResponse Text

The Amazon Resource Name (ARN) of the transform job.