Network.AWS.SageMaker.CreateTrial

Description

Creates an SageMaker trial. A trial is a set of steps called /trial components/ that produce a machine learning model. A trial is part of a single SageMaker experiment.

When you use SageMaker Studio or the SageMaker Python SDK, all experiments, trials, and trial components are automatically tracked, logged, and indexed. When you use the Amazon Web Services SDK for Python (Boto), you must use the logging APIs provided by the SDK.

You can add tags to a trial and then use the Search API to search for the tags.

To get a list of all your trials, call the ListTrials API. To view a trial's properties, call the DescribeTrial API. To create a trial component, call the CreateTrialComponent API.

Creating a Request

data CreateTrial

See: newCreateTrial smart constructor.

Constructors

CreateTrial'

Fields

metadataProperties :: Maybe MetadataProperties

displayName :: Maybe Text

The name of the trial as displayed. The name doesn't need to be unique. If DisplayName isn't specified, TrialName is displayed.

tags :: Maybe [Tag]

A list of tags to associate with the trial. You can use Search API to search on the tags.

trialName :: Text

The name of the trial. The name must be unique in your Amazon Web Services account and is not case-sensitive.

experimentName :: Text

The name of the experiment to associate the trial with.

newCreateTrial

Arguments

:: Text
$sel:trialName:CreateTrial'

-> Text
$sel:experimentName:CreateTrial'

-> CreateTrial

Create a value of CreateTrial with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metadataProperties:CreateTrial', createTrial_metadataProperties - Undocumented member.

$sel:displayName:CreateTrial', createTrial_displayName - The name of the trial as displayed. The name doesn't need to be unique. If DisplayName isn't specified, TrialName is displayed.

$sel:tags:CreateTrial', createTrial_tags - A list of tags to associate with the trial. You can use Search API to search on the tags.

$sel:trialName:CreateTrial', createTrial_trialName - The name of the trial. The name must be unique in your Amazon Web Services account and is not case-sensitive.

$sel:experimentName:CreateTrial', createTrial_experimentName - The name of the experiment to associate the trial with.

Request Lenses

createTrial_metadataProperties :: Lens' CreateTrial (Maybe MetadataProperties)

Undocumented member.

createTrial_displayName :: Lens' CreateTrial (Maybe Text)

The name of the trial as displayed. The name doesn't need to be unique. If DisplayName isn't specified, TrialName is displayed.

createTrial_tags :: Lens' CreateTrial (Maybe [Tag])

A list of tags to associate with the trial. You can use Search API to search on the tags.

createTrial_trialName :: Lens' CreateTrial Text

The name of the trial. The name must be unique in your Amazon Web Services account and is not case-sensitive.

createTrial_experimentName :: Lens' CreateTrial Text

The name of the experiment to associate the trial with.

Destructuring the Response

data CreateTrialResponse

See: newCreateTrialResponse smart constructor.

Constructors

CreateTrialResponse'

Fields

trialArn :: Maybe Text

The Amazon Resource Name (ARN) of the trial.

httpStatus :: Int

The response's http status code.

newCreateTrialResponse

Arguments

:: Int
$sel:httpStatus:CreateTrialResponse'

-> CreateTrialResponse

Create a value of CreateTrialResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trialArn:CreateTrialResponse', createTrialResponse_trialArn - The Amazon Resource Name (ARN) of the trial.

$sel:httpStatus:CreateTrialResponse', createTrialResponse_httpStatus - The response's http status code.

Response Lenses

createTrialResponse_trialArn :: Lens' CreateTrialResponse (Maybe Text)

The Amazon Resource Name (ARN) of the trial.

createTrialResponse_httpStatus :: Lens' CreateTrialResponse Int

The response's http status code.