Network.AWS.SageMaker.CreateTrialComponent

Description

Creates a trial component, which is a stage of a machine learning trial. A trial is composed of one or more trial components. A trial component can be used in multiple trials.

Trial components include pre-processing jobs, training jobs, and batch transform jobs.

When you use SageMaker Studio or the SageMaker Python SDK, all experiments, trials, and trial components are automatically tracked, logged, and indexed. When you use the Amazon Web Services SDK for Python (Boto), you must use the logging APIs provided by the SDK.

You can add tags to a trial component and then use the Search API to search for the tags.

Creating a Request

data CreateTrialComponent

See: newCreateTrialComponent smart constructor.

Constructors

CreateTrialComponent'

Fields

status :: Maybe TrialComponentStatus

The status of the component. States include:

InProgress

Completed

Failed

metadataProperties :: Maybe MetadataProperties

startTime :: Maybe POSIX

When the component started.

endTime :: Maybe POSIX

When the component ended.

tags :: Maybe [Tag]

A list of tags to associate with the component. You can use Search API to search on the tags.

inputArtifacts :: Maybe (HashMap Text TrialComponentArtifact)

The input artifacts for the component. Examples of input artifacts are datasets, algorithms, hyperparameters, source code, and instance types.

displayName :: Maybe Text

The name of the component as displayed. The name doesn't need to be unique. If DisplayName isn't specified, TrialComponentName is displayed.

outputArtifacts :: Maybe (HashMap Text TrialComponentArtifact)

The output artifacts for the component. Examples of output artifacts are metrics, snapshots, logs, and images.

parameters :: Maybe (HashMap Text TrialComponentParameterValue)

The hyperparameters for the component.

trialComponentName :: Text

The name of the component. The name must be unique in your Amazon Web Services account and is not case-sensitive.

newCreateTrialComponent

Arguments

:: Text
$sel:trialComponentName:CreateTrialComponent'

-> CreateTrialComponent

Create a value of CreateTrialComponent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateTrialComponent', createTrialComponent_status - The status of the component. States include:

InProgress

Completed

Failed

$sel:metadataProperties:CreateTrialComponent', createTrialComponent_metadataProperties - Undocumented member.

$sel:startTime:CreateTrialComponent', createTrialComponent_startTime - When the component started.

$sel:endTime:CreateTrialComponent', createTrialComponent_endTime - When the component ended.

$sel:tags:CreateTrialComponent', createTrialComponent_tags - A list of tags to associate with the component. You can use Search API to search on the tags.

$sel:inputArtifacts:CreateTrialComponent', createTrialComponent_inputArtifacts - The input artifacts for the component. Examples of input artifacts are datasets, algorithms, hyperparameters, source code, and instance types.

$sel:displayName:CreateTrialComponent', createTrialComponent_displayName - The name of the component as displayed. The name doesn't need to be unique. If DisplayName isn't specified, TrialComponentName is displayed.

$sel:outputArtifacts:CreateTrialComponent', createTrialComponent_outputArtifacts - The output artifacts for the component. Examples of output artifacts are metrics, snapshots, logs, and images.

$sel:parameters:CreateTrialComponent', createTrialComponent_parameters - The hyperparameters for the component.

$sel:trialComponentName:CreateTrialComponent', createTrialComponent_trialComponentName - The name of the component. The name must be unique in your Amazon Web Services account and is not case-sensitive.

Request Lenses

createTrialComponent_status :: Lens' CreateTrialComponent (Maybe TrialComponentStatus)

The status of the component. States include:

InProgress

Completed

Failed

createTrialComponent_metadataProperties :: Lens' CreateTrialComponent (Maybe MetadataProperties)

Undocumented member.

createTrialComponent_startTime :: Lens' CreateTrialComponent (Maybe UTCTime)

When the component started.

createTrialComponent_endTime :: Lens' CreateTrialComponent (Maybe UTCTime)

When the component ended.

createTrialComponent_tags :: Lens' CreateTrialComponent (Maybe [Tag])

A list of tags to associate with the component. You can use Search API to search on the tags.

createTrialComponent_inputArtifacts :: Lens' CreateTrialComponent (Maybe (HashMap Text TrialComponentArtifact))

The input artifacts for the component. Examples of input artifacts are datasets, algorithms, hyperparameters, source code, and instance types.

createTrialComponent_displayName :: Lens' CreateTrialComponent (Maybe Text)

The name of the component as displayed. The name doesn't need to be unique. If DisplayName isn't specified, TrialComponentName is displayed.

createTrialComponent_outputArtifacts :: Lens' CreateTrialComponent (Maybe (HashMap Text TrialComponentArtifact))

The output artifacts for the component. Examples of output artifacts are metrics, snapshots, logs, and images.

createTrialComponent_parameters :: Lens' CreateTrialComponent (Maybe (HashMap Text TrialComponentParameterValue))

The hyperparameters for the component.

createTrialComponent_trialComponentName :: Lens' CreateTrialComponent Text

The name of the component. The name must be unique in your Amazon Web Services account and is not case-sensitive.

Destructuring the Response

data CreateTrialComponentResponse

See: newCreateTrialComponentResponse smart constructor.

Constructors

CreateTrialComponentResponse'

Fields

trialComponentArn :: Maybe Text

The Amazon Resource Name (ARN) of the trial component.

httpStatus :: Int

The response's http status code.

newCreateTrialComponentResponse

Arguments

:: Int
$sel:httpStatus:CreateTrialComponentResponse'

-> CreateTrialComponentResponse

Create a value of CreateTrialComponentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trialComponentArn:CreateTrialComponentResponse', createTrialComponentResponse_trialComponentArn - The Amazon Resource Name (ARN) of the trial component.

$sel:httpStatus:CreateTrialComponentResponse', createTrialComponentResponse_httpStatus - The response's http status code.

Response Lenses

createTrialComponentResponse_trialComponentArn :: Lens' CreateTrialComponentResponse (Maybe Text)

The Amazon Resource Name (ARN) of the trial component.

createTrialComponentResponse_httpStatus :: Lens' CreateTrialComponentResponse Int

The response's http status code.