Network.AWS.SageMaker.CreateUserProfile

Description

Creates a user profile. A user profile represents a single user within a domain, and is the main way to reference a "person" for the purposes of sharing, reporting, and other user-oriented features. This entity is created when a user onboards to Amazon SageMaker Studio. If an administrator invites a person by email or imports them from SSO, a user profile is automatically created. A user profile is the primary holder of settings for an individual user and has a reference to the user's private Amazon Elastic File System (EFS) home directory.

Creating a Request

data CreateUserProfile

See: newCreateUserProfile smart constructor.

Constructors

CreateUserProfile'

Fields

userSettings :: Maybe UserSettings

A collection of settings.

singleSignOnUserValue :: Maybe Text

The username of the associated Amazon Web Services Single Sign-On User for this UserProfile. If the Domain's AuthMode is SSO, this field is required, and must match a valid username of a user in your directory. If the Domain's AuthMode is not SSO, this field cannot be specified.

singleSignOnUserIdentifier :: Maybe Text

A specifier for the type of value specified in SingleSignOnUserValue. Currently, the only supported value is "UserName". If the Domain's AuthMode is SSO, this field is required. If the Domain's AuthMode is not SSO, this field cannot be specified.

tags :: Maybe [Tag]

Each tag consists of a key and an optional value. Tag keys must be unique per resource.

Tags that you specify for the User Profile are also added to all Apps that the User Profile launches.

domainId :: Text

The ID of the associated Domain.

userProfileName :: Text

A name for the UserProfile. This value is not case sensitive.

newCreateUserProfile

Arguments

:: Text
$sel:domainId:CreateUserProfile'

-> Text
$sel:userProfileName:CreateUserProfile'

-> CreateUserProfile

Create a value of CreateUserProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userSettings:CreateUserProfile', createUserProfile_userSettings - A collection of settings.

$sel:singleSignOnUserValue:CreateUserProfile', createUserProfile_singleSignOnUserValue - The username of the associated Amazon Web Services Single Sign-On User for this UserProfile. If the Domain's AuthMode is SSO, this field is required, and must match a valid username of a user in your directory. If the Domain's AuthMode is not SSO, this field cannot be specified.

$sel:singleSignOnUserIdentifier:CreateUserProfile', createUserProfile_singleSignOnUserIdentifier - A specifier for the type of value specified in SingleSignOnUserValue. Currently, the only supported value is "UserName". If the Domain's AuthMode is SSO, this field is required. If the Domain's AuthMode is not SSO, this field cannot be specified.

$sel:tags:CreateUserProfile', createUserProfile_tags - Each tag consists of a key and an optional value. Tag keys must be unique per resource.

Tags that you specify for the User Profile are also added to all Apps that the User Profile launches.

$sel:domainId:CreateUserProfile', createUserProfile_domainId - The ID of the associated Domain.

$sel:userProfileName:CreateUserProfile', createUserProfile_userProfileName - A name for the UserProfile. This value is not case sensitive.

Request Lenses

createUserProfile_userSettings :: Lens' CreateUserProfile (Maybe UserSettings)

A collection of settings.

createUserProfile_singleSignOnUserValue :: Lens' CreateUserProfile (Maybe Text)

The username of the associated Amazon Web Services Single Sign-On User for this UserProfile. If the Domain's AuthMode is SSO, this field is required, and must match a valid username of a user in your directory. If the Domain's AuthMode is not SSO, this field cannot be specified.

createUserProfile_singleSignOnUserIdentifier :: Lens' CreateUserProfile (Maybe Text)

A specifier for the type of value specified in SingleSignOnUserValue. Currently, the only supported value is "UserName". If the Domain's AuthMode is SSO, this field is required. If the Domain's AuthMode is not SSO, this field cannot be specified.

createUserProfile_tags :: Lens' CreateUserProfile (Maybe [Tag])

Each tag consists of a key and an optional value. Tag keys must be unique per resource.

Tags that you specify for the User Profile are also added to all Apps that the User Profile launches.

createUserProfile_domainId :: Lens' CreateUserProfile Text

The ID of the associated Domain.

createUserProfile_userProfileName :: Lens' CreateUserProfile Text

A name for the UserProfile. This value is not case sensitive.

Destructuring the Response

data CreateUserProfileResponse

See: newCreateUserProfileResponse smart constructor.

Constructors

CreateUserProfileResponse'

Fields

userProfileArn :: Maybe Text

The user profile Amazon Resource Name (ARN).

httpStatus :: Int

The response's http status code.

newCreateUserProfileResponse

Arguments

:: Int
$sel:httpStatus:CreateUserProfileResponse'

-> CreateUserProfileResponse

Create a value of CreateUserProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userProfileArn:CreateUserProfileResponse', createUserProfileResponse_userProfileArn - The user profile Amazon Resource Name (ARN).

$sel:httpStatus:CreateUserProfileResponse', createUserProfileResponse_httpStatus - The response's http status code.

Response Lenses

createUserProfileResponse_userProfileArn :: Lens' CreateUserProfileResponse (Maybe Text)

The user profile Amazon Resource Name (ARN).

createUserProfileResponse_httpStatus :: Lens' CreateUserProfileResponse Int

The response's http status code.