Network.AWS.SageMaker.CreateWorkforce

Description

Use this operation to create a workforce. This operation will return an error if a workforce already exists in the Amazon Web Services Region that you specify. You can only create one workforce in each Amazon Web Services Region per Amazon Web Services account.

If you want to create a new workforce in an Amazon Web Services Region where a workforce already exists, use the API operation to delete the existing workforce and then use CreateWorkforce to create a new workforce.

To create a private workforce using Amazon Cognito, you must specify a Cognito user pool in CognitoConfig. You can also create an Amazon Cognito workforce using the Amazon SageMaker console. For more information, see Create a Private Workforce (Amazon Cognito).

To create a private workforce using your own OIDC Identity Provider (IdP), specify your IdP configuration in OidcConfig. Your OIDC IdP must support groups because groups are used by Ground Truth and Amazon A2I to create work teams. For more information, see Create a Private Workforce (OIDC IdP).

Creating a Request

data CreateWorkforce

See: newCreateWorkforce smart constructor.

Constructors

CreateWorkforce'

Fields

tags :: Maybe [Tag]

An array of key-value pairs that contain metadata to help you categorize and organize our workforce. Each tag consists of a key and a value, both of which you define.

sourceIpConfig :: Maybe SourceIpConfig

oidcConfig :: Maybe OidcConfig

Use this parameter to configure a private workforce using your own OIDC Identity Provider.

Do not use CognitoConfig if you specify values for OidcConfig.

cognitoConfig :: Maybe CognitoConfig

Use this parameter to configure an Amazon Cognito private workforce. A single Cognito workforce is created using and corresponds to a single Amazon Cognito user pool.

Do not use OidcConfig if you specify values for CognitoConfig.

workforceName :: Text

The name of the private workforce.

newCreateWorkforce

Arguments

:: Text
$sel:workforceName:CreateWorkforce'

-> CreateWorkforce

Create a value of CreateWorkforce with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateWorkforce', createWorkforce_tags - An array of key-value pairs that contain metadata to help you categorize and organize our workforce. Each tag consists of a key and a value, both of which you define.

$sel:sourceIpConfig:CreateWorkforce', createWorkforce_sourceIpConfig - Undocumented member.

$sel:oidcConfig:CreateWorkforce', createWorkforce_oidcConfig - Use this parameter to configure a private workforce using your own OIDC Identity Provider.

Do not use CognitoConfig if you specify values for OidcConfig.

$sel:cognitoConfig:CreateWorkforce', createWorkforce_cognitoConfig - Use this parameter to configure an Amazon Cognito private workforce. A single Cognito workforce is created using and corresponds to a single Amazon Cognito user pool.

Do not use OidcConfig if you specify values for CognitoConfig.

$sel:workforceName:CreateWorkforce', createWorkforce_workforceName - The name of the private workforce.

Request Lenses

createWorkforce_tags :: Lens' CreateWorkforce (Maybe [Tag])

An array of key-value pairs that contain metadata to help you categorize and organize our workforce. Each tag consists of a key and a value, both of which you define.

createWorkforce_sourceIpConfig :: Lens' CreateWorkforce (Maybe SourceIpConfig)

Undocumented member.

createWorkforce_oidcConfig :: Lens' CreateWorkforce (Maybe OidcConfig)

Use this parameter to configure a private workforce using your own OIDC Identity Provider.

Do not use CognitoConfig if you specify values for OidcConfig.

createWorkforce_cognitoConfig :: Lens' CreateWorkforce (Maybe CognitoConfig)

Use this parameter to configure an Amazon Cognito private workforce. A single Cognito workforce is created using and corresponds to a single Amazon Cognito user pool.

Do not use OidcConfig if you specify values for CognitoConfig.

createWorkforce_workforceName :: Lens' CreateWorkforce Text

The name of the private workforce.

Destructuring the Response

data CreateWorkforceResponse

See: newCreateWorkforceResponse smart constructor.

Constructors

CreateWorkforceResponse'

Fields

httpStatus :: Int

The response's http status code.

workforceArn :: Text

The Amazon Resource Name (ARN) of the workforce.

newCreateWorkforceResponse

Arguments

:: Int
$sel:httpStatus:CreateWorkforceResponse'

-> Text
$sel:workforceArn:CreateWorkforceResponse'

-> CreateWorkforceResponse

Create a value of CreateWorkforceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateWorkforceResponse', createWorkforceResponse_httpStatus - The response's http status code.

$sel:workforceArn:CreateWorkforceResponse', createWorkforceResponse_workforceArn - The Amazon Resource Name (ARN) of the workforce.

Response Lenses

createWorkforceResponse_httpStatus :: Lens' CreateWorkforceResponse Int

The response's http status code.

createWorkforceResponse_workforceArn :: Lens' CreateWorkforceResponse Text

The Amazon Resource Name (ARN) of the workforce.