Network.AWS.SageMaker.CreateWorkteam

Description

Creates a new work team for labeling your data. A work team is defined by one or more Amazon Cognito user pools. You must first create the user pools before you can create a work team.

You cannot create more than 25 work teams in an account and region.

Creating a Request

data CreateWorkteam

See: newCreateWorkteam smart constructor.

Constructors

CreateWorkteam'

Fields

workforceName :: Maybe Text

The name of the workforce.

notificationConfiguration :: Maybe NotificationConfiguration

Configures notification of workers regarding available or expiring work items.

tags :: Maybe [Tag]

An array of key-value pairs.

For more information, see Resource Tag and Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

workteamName :: Text

The name of the work team. Use this name to identify the work team.

memberDefinitions :: NonEmpty MemberDefinition

A list of MemberDefinition objects that contains objects that identify the workers that make up the work team.

Workforces can be created using Amazon Cognito or your own OIDC Identity Provider (IdP). For private workforces created using Amazon Cognito use CognitoMemberDefinition. For workforces created using your own OIDC identity provider (IdP) use OidcMemberDefinition. Do not provide input for both of these parameters in a single request.

For workforces created using Amazon Cognito, private work teams correspond to Amazon Cognito user groups within the user pool used to create a workforce. All of the CognitoMemberDefinition objects that make up the member definition must have the same ClientId and UserPool values. To add a Amazon Cognito user group to an existing worker pool, see groups to a User Pool. For more information about user pools, see Amazon Cognito User Pools.

For workforces created using your own OIDC IdP, specify the user groups that you want to include in your private work team in OidcMemberDefinition by listing those groups in Groups.

description :: Text

A description of the work team.

newCreateWorkteam

Arguments

:: Text
$sel:workteamName:CreateWorkteam'

-> NonEmpty MemberDefinition
$sel:memberDefinitions:CreateWorkteam'

-> Text
$sel:description:CreateWorkteam'

-> CreateWorkteam

Create a value of CreateWorkteam with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workforceName:CreateWorkteam', createWorkteam_workforceName - The name of the workforce.

$sel:notificationConfiguration:CreateWorkteam', createWorkteam_notificationConfiguration - Configures notification of workers regarding available or expiring work items.

$sel:tags:CreateWorkteam', createWorkteam_tags - An array of key-value pairs.

For more information, see Resource Tag and Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

$sel:workteamName:CreateWorkteam', createWorkteam_workteamName - The name of the work team. Use this name to identify the work team.

$sel:memberDefinitions:CreateWorkteam', createWorkteam_memberDefinitions - A list of MemberDefinition objects that contains objects that identify the workers that make up the work team.

Workforces can be created using Amazon Cognito or your own OIDC Identity Provider (IdP). For private workforces created using Amazon Cognito use CognitoMemberDefinition. For workforces created using your own OIDC identity provider (IdP) use OidcMemberDefinition. Do not provide input for both of these parameters in a single request.

For workforces created using Amazon Cognito, private work teams correspond to Amazon Cognito user groups within the user pool used to create a workforce. All of the CognitoMemberDefinition objects that make up the member definition must have the same ClientId and UserPool values. To add a Amazon Cognito user group to an existing worker pool, see groups to a User Pool. For more information about user pools, see Amazon Cognito User Pools.

For workforces created using your own OIDC IdP, specify the user groups that you want to include in your private work team in OidcMemberDefinition by listing those groups in Groups.

$sel:description:CreateWorkteam', createWorkteam_description - A description of the work team.

Request Lenses

createWorkteam_workforceName :: Lens' CreateWorkteam (Maybe Text)

The name of the workforce.

createWorkteam_notificationConfiguration :: Lens' CreateWorkteam (Maybe NotificationConfiguration)

Configures notification of workers regarding available or expiring work items.

createWorkteam_tags :: Lens' CreateWorkteam (Maybe [Tag])

An array of key-value pairs.

For more information, see Resource Tag and Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

createWorkteam_workteamName :: Lens' CreateWorkteam Text

The name of the work team. Use this name to identify the work team.

createWorkteam_memberDefinitions :: Lens' CreateWorkteam (NonEmpty MemberDefinition)

A list of MemberDefinition objects that contains objects that identify the workers that make up the work team.

Workforces can be created using Amazon Cognito or your own OIDC Identity Provider (IdP). For private workforces created using Amazon Cognito use CognitoMemberDefinition. For workforces created using your own OIDC identity provider (IdP) use OidcMemberDefinition. Do not provide input for both of these parameters in a single request.

For workforces created using Amazon Cognito, private work teams correspond to Amazon Cognito user groups within the user pool used to create a workforce. All of the CognitoMemberDefinition objects that make up the member definition must have the same ClientId and UserPool values. To add a Amazon Cognito user group to an existing worker pool, see groups to a User Pool. For more information about user pools, see Amazon Cognito User Pools.

For workforces created using your own OIDC IdP, specify the user groups that you want to include in your private work team in OidcMemberDefinition by listing those groups in Groups.

createWorkteam_description :: Lens' CreateWorkteam Text

A description of the work team.

Destructuring the Response

data CreateWorkteamResponse

See: newCreateWorkteamResponse smart constructor.

Constructors

CreateWorkteamResponse'

Fields

workteamArn :: Maybe Text

The Amazon Resource Name (ARN) of the work team. You can use this ARN to identify the work team.

httpStatus :: Int

The response's http status code.

newCreateWorkteamResponse

Arguments

:: Int
$sel:httpStatus:CreateWorkteamResponse'

-> CreateWorkteamResponse

Create a value of CreateWorkteamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workteamArn:CreateWorkteamResponse', createWorkteamResponse_workteamArn - The Amazon Resource Name (ARN) of the work team. You can use this ARN to identify the work team.

$sel:httpStatus:CreateWorkteamResponse', createWorkteamResponse_httpStatus - The response's http status code.

Response Lenses

createWorkteamResponse_workteamArn :: Lens' CreateWorkteamResponse (Maybe Text)

The Amazon Resource Name (ARN) of the work team. You can use this ARN to identify the work team.

createWorkteamResponse_httpStatus :: Lens' CreateWorkteamResponse Int

The response's http status code.