Network.AWS.SageMaker.DeleteArtifact

Description

Deletes an artifact. Either ArtifactArn or Source must be specified.

Creating a Request

data DeleteArtifact

See: newDeleteArtifact smart constructor.

Constructors

DeleteArtifact'

Fields

artifactArn :: Maybe Text

The Amazon Resource Name (ARN) of the artifact to delete.

source :: Maybe ArtifactSource

The URI of the source.

newDeleteArtifact :: DeleteArtifact

Create a value of DeleteArtifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:artifactArn:DeleteArtifact', deleteArtifact_artifactArn - The Amazon Resource Name (ARN) of the artifact to delete.

$sel:source:DeleteArtifact', deleteArtifact_source - The URI of the source.

Request Lenses

deleteArtifact_artifactArn :: Lens' DeleteArtifact (Maybe Text)

The Amazon Resource Name (ARN) of the artifact to delete.

deleteArtifact_source :: Lens' DeleteArtifact (Maybe ArtifactSource)

The URI of the source.

Destructuring the Response

data DeleteArtifactResponse

See: newDeleteArtifactResponse smart constructor.

Constructors

DeleteArtifactResponse'

Fields

artifactArn :: Maybe Text

The Amazon Resource Name (ARN) of the artifact.

httpStatus :: Int

The response's http status code.

newDeleteArtifactResponse

Arguments

:: Int
$sel:httpStatus:DeleteArtifactResponse'

-> DeleteArtifactResponse

Create a value of DeleteArtifactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:artifactArn:DeleteArtifact', deleteArtifactResponse_artifactArn - The Amazon Resource Name (ARN) of the artifact.

$sel:httpStatus:DeleteArtifactResponse', deleteArtifactResponse_httpStatus - The response's http status code.

Response Lenses

deleteArtifactResponse_artifactArn :: Lens' DeleteArtifactResponse (Maybe Text)

The Amazon Resource Name (ARN) of the artifact.

deleteArtifactResponse_httpStatus :: Lens' DeleteArtifactResponse Int

The response's http status code.