Network.AWS.SageMaker.DescribeAlgorithm

Description

Returns a description of the specified algorithm that is in your account.

Creating a Request

data DescribeAlgorithm

See: newDescribeAlgorithm smart constructor.

Constructors

DescribeAlgorithm'

Fields

algorithmName :: Text

The name of the algorithm to describe.

newDescribeAlgorithm

Arguments

:: Text
$sel:algorithmName:DescribeAlgorithm'

-> DescribeAlgorithm

Create a value of DescribeAlgorithm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:algorithmName:DescribeAlgorithm', describeAlgorithm_algorithmName - The name of the algorithm to describe.

Request Lenses

describeAlgorithm_algorithmName :: Lens' DescribeAlgorithm Text

The name of the algorithm to describe.

Destructuring the Response

data DescribeAlgorithmResponse

See: newDescribeAlgorithmResponse smart constructor.

Constructors

DescribeAlgorithmResponse'

Fields

validationSpecification :: Maybe AlgorithmValidationSpecification

Details about configurations for one or more training jobs that Amazon SageMaker runs to test the algorithm.

inferenceSpecification :: Maybe InferenceSpecification

Details about inference jobs that the algorithm runs.

algorithmDescription :: Maybe Text

A brief summary about the algorithm.

certifyForMarketplace :: Maybe Bool

Whether the algorithm is certified to be listed in Amazon Web Services Marketplace.

productId :: Maybe Text

The product identifier of the algorithm.

httpStatus :: Int

The response's http status code.

algorithmName :: Text

The name of the algorithm being described.

algorithmArn :: Text

The Amazon Resource Name (ARN) of the algorithm.

creationTime :: POSIX

A timestamp specifying when the algorithm was created.

trainingSpecification :: TrainingSpecification

Details about training jobs run by this algorithm.

algorithmStatus :: AlgorithmStatus

The current status of the algorithm.

algorithmStatusDetails :: AlgorithmStatusDetails

Details about the current status of the algorithm.

newDescribeAlgorithmResponse

Arguments

:: Int
$sel:httpStatus:DescribeAlgorithmResponse'

-> Text
$sel:algorithmName:DescribeAlgorithm'

-> Text
$sel:algorithmArn:DescribeAlgorithmResponse'

-> UTCTime
$sel:creationTime:DescribeAlgorithmResponse'

-> TrainingSpecification
$sel:trainingSpecification:DescribeAlgorithmResponse'

-> AlgorithmStatus
$sel:algorithmStatus:DescribeAlgorithmResponse'

-> AlgorithmStatusDetails
$sel:algorithmStatusDetails:DescribeAlgorithmResponse'

-> DescribeAlgorithmResponse

Create a value of DescribeAlgorithmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:validationSpecification:DescribeAlgorithmResponse', describeAlgorithmResponse_validationSpecification - Details about configurations for one or more training jobs that Amazon SageMaker runs to test the algorithm.

$sel:inferenceSpecification:DescribeAlgorithmResponse', describeAlgorithmResponse_inferenceSpecification - Details about inference jobs that the algorithm runs.

$sel:algorithmDescription:DescribeAlgorithmResponse', describeAlgorithmResponse_algorithmDescription - A brief summary about the algorithm.

$sel:certifyForMarketplace:DescribeAlgorithmResponse', describeAlgorithmResponse_certifyForMarketplace - Whether the algorithm is certified to be listed in Amazon Web Services Marketplace.

$sel:productId:DescribeAlgorithmResponse', describeAlgorithmResponse_productId - The product identifier of the algorithm.

$sel:httpStatus:DescribeAlgorithmResponse', describeAlgorithmResponse_httpStatus - The response's http status code.

$sel:algorithmName:DescribeAlgorithm', describeAlgorithmResponse_algorithmName - The name of the algorithm being described.

$sel:algorithmArn:DescribeAlgorithmResponse', describeAlgorithmResponse_algorithmArn - The Amazon Resource Name (ARN) of the algorithm.

$sel:creationTime:DescribeAlgorithmResponse', describeAlgorithmResponse_creationTime - A timestamp specifying when the algorithm was created.

$sel:trainingSpecification:DescribeAlgorithmResponse', describeAlgorithmResponse_trainingSpecification - Details about training jobs run by this algorithm.

$sel:algorithmStatus:DescribeAlgorithmResponse', describeAlgorithmResponse_algorithmStatus - The current status of the algorithm.

$sel:algorithmStatusDetails:DescribeAlgorithmResponse', describeAlgorithmResponse_algorithmStatusDetails - Details about the current status of the algorithm.

Response Lenses

describeAlgorithmResponse_validationSpecification :: Lens' DescribeAlgorithmResponse (Maybe AlgorithmValidationSpecification)

Details about configurations for one or more training jobs that Amazon SageMaker runs to test the algorithm.

describeAlgorithmResponse_inferenceSpecification :: Lens' DescribeAlgorithmResponse (Maybe InferenceSpecification)

Details about inference jobs that the algorithm runs.

describeAlgorithmResponse_algorithmDescription :: Lens' DescribeAlgorithmResponse (Maybe Text)

A brief summary about the algorithm.

describeAlgorithmResponse_certifyForMarketplace :: Lens' DescribeAlgorithmResponse (Maybe Bool)

Whether the algorithm is certified to be listed in Amazon Web Services Marketplace.

describeAlgorithmResponse_productId :: Lens' DescribeAlgorithmResponse (Maybe Text)

The product identifier of the algorithm.

describeAlgorithmResponse_httpStatus :: Lens' DescribeAlgorithmResponse Int

The response's http status code.

describeAlgorithmResponse_algorithmName :: Lens' DescribeAlgorithmResponse Text

The name of the algorithm being described.

describeAlgorithmResponse_algorithmArn :: Lens' DescribeAlgorithmResponse Text

The Amazon Resource Name (ARN) of the algorithm.

describeAlgorithmResponse_creationTime :: Lens' DescribeAlgorithmResponse UTCTime

A timestamp specifying when the algorithm was created.

describeAlgorithmResponse_trainingSpecification :: Lens' DescribeAlgorithmResponse TrainingSpecification

Details about training jobs run by this algorithm.

describeAlgorithmResponse_algorithmStatus :: Lens' DescribeAlgorithmResponse AlgorithmStatus

The current status of the algorithm.

describeAlgorithmResponse_algorithmStatusDetails :: Lens' DescribeAlgorithmResponse AlgorithmStatusDetails

Details about the current status of the algorithm.