Network.AWS.SageMaker.DescribeApp

Description

Describes the app.

Creating a Request

data DescribeApp

See: newDescribeApp smart constructor.

Constructors

DescribeApp'

Fields

domainId :: Text

The domain ID.

userProfileName :: Text

The user profile name.

appType :: AppType

The type of app.

appName :: Text

The name of the app.

newDescribeApp

Arguments

:: Text
$sel:domainId:DescribeApp'

-> Text
$sel:userProfileName:DescribeApp'

-> AppType
$sel:appType:DescribeApp'

-> Text
$sel:appName:DescribeApp'

-> DescribeApp

Create a value of DescribeApp with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainId:DescribeApp', describeApp_domainId - The domain ID.

$sel:userProfileName:DescribeApp', describeApp_userProfileName - The user profile name.

$sel:appType:DescribeApp', describeApp_appType - The type of app.

$sel:appName:DescribeApp', describeApp_appName - The name of the app.

Request Lenses

describeApp_domainId :: Lens' DescribeApp Text

The domain ID.

describeApp_userProfileName :: Lens' DescribeApp Text

The user profile name.

describeApp_appType :: Lens' DescribeApp AppType

The type of app.

describeApp_appName :: Lens' DescribeApp Text

The name of the app.

Destructuring the Response

data DescribeAppResponse

See: newDescribeAppResponse smart constructor.

Constructors

DescribeAppResponse'

Fields

resourceSpec :: Maybe ResourceSpec

The instance type and the Amazon Resource Name (ARN) of the SageMaker image created on the instance.

creationTime :: Maybe POSIX

The creation time.

status :: Maybe AppStatus

The status.

appType :: Maybe AppType

The type of app.

appName :: Maybe Text

The name of the app.

userProfileName :: Maybe Text

The user profile name.

domainId :: Maybe Text

The domain ID.

appArn :: Maybe Text

The Amazon Resource Name (ARN) of the app.

failureReason :: Maybe Text

The failure reason.

lastHealthCheckTimestamp :: Maybe POSIX

The timestamp of the last health check.

lastUserActivityTimestamp :: Maybe POSIX

The timestamp of the last user's activity.

httpStatus :: Int

The response's http status code.

newDescribeAppResponse

Arguments

:: Int
$sel:httpStatus:DescribeAppResponse'

-> DescribeAppResponse

Create a value of DescribeAppResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceSpec:DescribeAppResponse', describeAppResponse_resourceSpec - The instance type and the Amazon Resource Name (ARN) of the SageMaker image created on the instance.

$sel:creationTime:DescribeAppResponse', describeAppResponse_creationTime - The creation time.

$sel:status:DescribeAppResponse', describeAppResponse_status - The status.

$sel:appType:DescribeApp', describeAppResponse_appType - The type of app.

$sel:appName:DescribeApp', describeAppResponse_appName - The name of the app.

$sel:userProfileName:DescribeApp', describeAppResponse_userProfileName - The user profile name.

$sel:domainId:DescribeApp', describeAppResponse_domainId - The domain ID.

$sel:appArn:DescribeAppResponse', describeAppResponse_appArn - The Amazon Resource Name (ARN) of the app.

$sel:failureReason:DescribeAppResponse', describeAppResponse_failureReason - The failure reason.

$sel:lastHealthCheckTimestamp:DescribeAppResponse', describeAppResponse_lastHealthCheckTimestamp - The timestamp of the last health check.

$sel:lastUserActivityTimestamp:DescribeAppResponse', describeAppResponse_lastUserActivityTimestamp - The timestamp of the last user's activity.

$sel:httpStatus:DescribeAppResponse', describeAppResponse_httpStatus - The response's http status code.

Response Lenses

describeAppResponse_resourceSpec :: Lens' DescribeAppResponse (Maybe ResourceSpec)

The instance type and the Amazon Resource Name (ARN) of the SageMaker image created on the instance.

describeAppResponse_creationTime :: Lens' DescribeAppResponse (Maybe UTCTime)

The creation time.

describeAppResponse_status :: Lens' DescribeAppResponse (Maybe AppStatus)

The status.

describeAppResponse_appType :: Lens' DescribeAppResponse (Maybe AppType)

The type of app.

describeAppResponse_appName :: Lens' DescribeAppResponse (Maybe Text)

The name of the app.

describeAppResponse_userProfileName :: Lens' DescribeAppResponse (Maybe Text)

The user profile name.

describeAppResponse_domainId :: Lens' DescribeAppResponse (Maybe Text)

The domain ID.

describeAppResponse_appArn :: Lens' DescribeAppResponse (Maybe Text)

The Amazon Resource Name (ARN) of the app.

describeAppResponse_failureReason :: Lens' DescribeAppResponse (Maybe Text)

The failure reason.

describeAppResponse_lastHealthCheckTimestamp :: Lens' DescribeAppResponse (Maybe UTCTime)

The timestamp of the last health check.

describeAppResponse_lastUserActivityTimestamp :: Lens' DescribeAppResponse (Maybe UTCTime)

The timestamp of the last user's activity.

describeAppResponse_httpStatus :: Lens' DescribeAppResponse Int

The response's http status code.