Network.AWS.SageMaker.DescribeArtifact

Description

Describes an artifact.

Creating a Request

data DescribeArtifact

See: newDescribeArtifact smart constructor.

Constructors

DescribeArtifact'

Fields

artifactArn :: Text

The Amazon Resource Name (ARN) of the artifact to describe.

newDescribeArtifact

Arguments

:: Text
$sel:artifactArn:DescribeArtifact'

-> DescribeArtifact

Create a value of DescribeArtifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:artifactArn:DescribeArtifact', describeArtifact_artifactArn - The Amazon Resource Name (ARN) of the artifact to describe.

Request Lenses

describeArtifact_artifactArn :: Lens' DescribeArtifact Text

The Amazon Resource Name (ARN) of the artifact to describe.

Destructuring the Response

data DescribeArtifactResponse

See: newDescribeArtifactResponse smart constructor.

Constructors

DescribeArtifactResponse'

Fields

creationTime :: Maybe POSIX

When the artifact was created.

metadataProperties :: Maybe MetadataProperties

artifactName :: Maybe Text

The name of the artifact.

artifactType :: Maybe Text

The type of the artifact.

artifactArn :: Maybe Text

The Amazon Resource Name (ARN) of the artifact.

source :: Maybe ArtifactSource

The source of the artifact.

properties :: Maybe (HashMap Text Text)

A list of the artifact's properties.

lastModifiedTime :: Maybe POSIX

When the artifact was last modified.

createdBy :: Maybe UserContext

lastModifiedBy :: Maybe UserContext

httpStatus :: Int

The response's http status code.

newDescribeArtifactResponse

Arguments

:: Int
$sel:httpStatus:DescribeArtifactResponse'

-> DescribeArtifactResponse

Create a value of DescribeArtifactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeArtifactResponse', describeArtifactResponse_creationTime - When the artifact was created.

$sel:metadataProperties:DescribeArtifactResponse', describeArtifactResponse_metadataProperties - Undocumented member.

$sel:artifactName:DescribeArtifactResponse', describeArtifactResponse_artifactName - The name of the artifact.

$sel:artifactType:DescribeArtifactResponse', describeArtifactResponse_artifactType - The type of the artifact.

$sel:artifactArn:DescribeArtifact', describeArtifactResponse_artifactArn - The Amazon Resource Name (ARN) of the artifact.

$sel:source:DescribeArtifactResponse', describeArtifactResponse_source - The source of the artifact.

$sel:properties:DescribeArtifactResponse', describeArtifactResponse_properties - A list of the artifact's properties.

$sel:lastModifiedTime:DescribeArtifactResponse', describeArtifactResponse_lastModifiedTime - When the artifact was last modified.

$sel:createdBy:DescribeArtifactResponse', describeArtifactResponse_createdBy - Undocumented member.

$sel:lastModifiedBy:DescribeArtifactResponse', describeArtifactResponse_lastModifiedBy - Undocumented member.

$sel:httpStatus:DescribeArtifactResponse', describeArtifactResponse_httpStatus - The response's http status code.

Response Lenses

describeArtifactResponse_creationTime :: Lens' DescribeArtifactResponse (Maybe UTCTime)

When the artifact was created.

describeArtifactResponse_metadataProperties :: Lens' DescribeArtifactResponse (Maybe MetadataProperties)

Undocumented member.

describeArtifactResponse_artifactName :: Lens' DescribeArtifactResponse (Maybe Text)

The name of the artifact.

describeArtifactResponse_artifactType :: Lens' DescribeArtifactResponse (Maybe Text)

The type of the artifact.

describeArtifactResponse_artifactArn :: Lens' DescribeArtifactResponse (Maybe Text)

The Amazon Resource Name (ARN) of the artifact.

describeArtifactResponse_source :: Lens' DescribeArtifactResponse (Maybe ArtifactSource)

The source of the artifact.

describeArtifactResponse_properties :: Lens' DescribeArtifactResponse (Maybe (HashMap Text Text))

A list of the artifact's properties.

describeArtifactResponse_lastModifiedTime :: Lens' DescribeArtifactResponse (Maybe UTCTime)

When the artifact was last modified.

describeArtifactResponse_createdBy :: Lens' DescribeArtifactResponse (Maybe UserContext)

Undocumented member.

describeArtifactResponse_lastModifiedBy :: Lens' DescribeArtifactResponse (Maybe UserContext)

Undocumented member.

describeArtifactResponse_httpStatus :: Lens' DescribeArtifactResponse Int

The response's http status code.