Network.AWS.SageMaker.DescribeAutoMLJob

Description

Returns information about an Amazon SageMaker AutoML job.

Creating a Request

data DescribeAutoMLJob

See: newDescribeAutoMLJob smart constructor.

Constructors

DescribeAutoMLJob'

Fields

autoMLJobName :: Text

Requests information about an AutoML job using its unique name.

newDescribeAutoMLJob

Arguments

:: Text
$sel:autoMLJobName:DescribeAutoMLJob'

-> DescribeAutoMLJob

Create a value of DescribeAutoMLJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoMLJobName:DescribeAutoMLJob', describeAutoMLJob_autoMLJobName - Requests information about an AutoML job using its unique name.

Request Lenses

describeAutoMLJob_autoMLJobName :: Lens' DescribeAutoMLJob Text

Requests information about an AutoML job using its unique name.

Destructuring the Response

data DescribeAutoMLJobResponse

See: newDescribeAutoMLJobResponse smart constructor.

Constructors

DescribeAutoMLJobResponse'

Fields

generateCandidateDefinitionsOnly :: Maybe Bool

Indicates whether the output for an AutoML job generates candidate definitions only.

modelDeployResult :: Maybe ModelDeployResult

Provides information about endpoint for the model deployment.

endTime :: Maybe POSIX

Returns the end time of the AutoML job.

resolvedAttributes :: Maybe ResolvedAttributes

This contains ProblemType, AutoMLJobObjective, and CompletionCriteria. If you do not provide these values, they are auto-inferred. If you do provide them, the values used are the ones you provide.

autoMLJobArtifacts :: Maybe AutoMLJobArtifacts

Returns information on the job's artifacts found in AutoMLJobArtifacts.

partialFailureReasons :: Maybe (NonEmpty AutoMLPartialFailureReason)

Returns a list of reasons for partial failures within an AutoML job.

failureReason :: Maybe Text

Returns the failure reason for an AutoML job, when applicable.

autoMLJobObjective :: Maybe AutoMLJobObjective

Returns the job's objective.

modelDeployConfig :: Maybe ModelDeployConfig

Indicates whether the model was deployed automatically to an endpoint and the name of that endpoint if deployed automatically.

autoMLJobConfig :: Maybe AutoMLJobConfig

Returns the configuration for the AutoML job.

problemType :: Maybe ProblemType

Returns the job's problem type.

bestCandidate :: Maybe AutoMLCandidate

Returns the job's best AutoMLCandidate.

httpStatus :: Int

The response's http status code.

autoMLJobName :: Text

Returns the name of the AutoML job.

autoMLJobArn :: Text

Returns the ARN of the AutoML job.

inputDataConfig :: NonEmpty AutoMLChannel

Returns the input data configuration for the AutoML job..

outputDataConfig :: AutoMLOutputDataConfig

Returns the job's output data config.

roleArn :: Text

The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that has read permission to the input data location and write permission to the output data location in Amazon S3.

creationTime :: POSIX

Returns the creation time of the AutoML job.

lastModifiedTime :: POSIX

Returns the job's last modified time.

autoMLJobStatus :: AutoMLJobStatus

Returns the status of the AutoML job.

autoMLJobSecondaryStatus :: AutoMLJobSecondaryStatus

Returns the secondary status of the AutoML job.

newDescribeAutoMLJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeAutoMLJobResponse'

-> Text
$sel:autoMLJobName:DescribeAutoMLJob'

-> Text
$sel:autoMLJobArn:DescribeAutoMLJobResponse'

-> NonEmpty AutoMLChannel
$sel:inputDataConfig:DescribeAutoMLJobResponse'

-> AutoMLOutputDataConfig
$sel:outputDataConfig:DescribeAutoMLJobResponse'

-> Text
$sel:roleArn:DescribeAutoMLJobResponse'

-> UTCTime
$sel:creationTime:DescribeAutoMLJobResponse'

-> UTCTime
$sel:lastModifiedTime:DescribeAutoMLJobResponse'

-> AutoMLJobStatus
$sel:autoMLJobStatus:DescribeAutoMLJobResponse'

-> AutoMLJobSecondaryStatus
$sel:autoMLJobSecondaryStatus:DescribeAutoMLJobResponse'

-> DescribeAutoMLJobResponse

Create a value of DescribeAutoMLJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:generateCandidateDefinitionsOnly:DescribeAutoMLJobResponse', describeAutoMLJobResponse_generateCandidateDefinitionsOnly - Indicates whether the output for an AutoML job generates candidate definitions only.

$sel:modelDeployResult:DescribeAutoMLJobResponse', describeAutoMLJobResponse_modelDeployResult - Provides information about endpoint for the model deployment.

$sel:endTime:DescribeAutoMLJobResponse', describeAutoMLJobResponse_endTime - Returns the end time of the AutoML job.

$sel:resolvedAttributes:DescribeAutoMLJobResponse', describeAutoMLJobResponse_resolvedAttributes - This contains ProblemType, AutoMLJobObjective, and CompletionCriteria. If you do not provide these values, they are auto-inferred. If you do provide them, the values used are the ones you provide.

$sel:autoMLJobArtifacts:DescribeAutoMLJobResponse', describeAutoMLJobResponse_autoMLJobArtifacts - Returns information on the job's artifacts found in AutoMLJobArtifacts.

$sel:partialFailureReasons:DescribeAutoMLJobResponse', describeAutoMLJobResponse_partialFailureReasons - Returns a list of reasons for partial failures within an AutoML job.

$sel:failureReason:DescribeAutoMLJobResponse', describeAutoMLJobResponse_failureReason - Returns the failure reason for an AutoML job, when applicable.

$sel:autoMLJobObjective:DescribeAutoMLJobResponse', describeAutoMLJobResponse_autoMLJobObjective - Returns the job's objective.

$sel:modelDeployConfig:DescribeAutoMLJobResponse', describeAutoMLJobResponse_modelDeployConfig - Indicates whether the model was deployed automatically to an endpoint and the name of that endpoint if deployed automatically.

$sel:autoMLJobConfig:DescribeAutoMLJobResponse', describeAutoMLJobResponse_autoMLJobConfig - Returns the configuration for the AutoML job.

$sel:problemType:DescribeAutoMLJobResponse', describeAutoMLJobResponse_problemType - Returns the job's problem type.

$sel:bestCandidate:DescribeAutoMLJobResponse', describeAutoMLJobResponse_bestCandidate - Returns the job's best AutoMLCandidate.

$sel:httpStatus:DescribeAutoMLJobResponse', describeAutoMLJobResponse_httpStatus - The response's http status code.

$sel:autoMLJobName:DescribeAutoMLJob', describeAutoMLJobResponse_autoMLJobName - Returns the name of the AutoML job.

$sel:autoMLJobArn:DescribeAutoMLJobResponse', describeAutoMLJobResponse_autoMLJobArn - Returns the ARN of the AutoML job.

$sel:inputDataConfig:DescribeAutoMLJobResponse', describeAutoMLJobResponse_inputDataConfig - Returns the input data configuration for the AutoML job..

$sel:outputDataConfig:DescribeAutoMLJobResponse', describeAutoMLJobResponse_outputDataConfig - Returns the job's output data config.

$sel:roleArn:DescribeAutoMLJobResponse', describeAutoMLJobResponse_roleArn - The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that has read permission to the input data location and write permission to the output data location in Amazon S3.

$sel:creationTime:DescribeAutoMLJobResponse', describeAutoMLJobResponse_creationTime - Returns the creation time of the AutoML job.

$sel:lastModifiedTime:DescribeAutoMLJobResponse', describeAutoMLJobResponse_lastModifiedTime - Returns the job's last modified time.

$sel:autoMLJobStatus:DescribeAutoMLJobResponse', describeAutoMLJobResponse_autoMLJobStatus - Returns the status of the AutoML job.

$sel:autoMLJobSecondaryStatus:DescribeAutoMLJobResponse', describeAutoMLJobResponse_autoMLJobSecondaryStatus - Returns the secondary status of the AutoML job.

Response Lenses

describeAutoMLJobResponse_generateCandidateDefinitionsOnly :: Lens' DescribeAutoMLJobResponse (Maybe Bool)

Indicates whether the output for an AutoML job generates candidate definitions only.

describeAutoMLJobResponse_modelDeployResult :: Lens' DescribeAutoMLJobResponse (Maybe ModelDeployResult)

Provides information about endpoint for the model deployment.

describeAutoMLJobResponse_endTime :: Lens' DescribeAutoMLJobResponse (Maybe UTCTime)

Returns the end time of the AutoML job.

describeAutoMLJobResponse_resolvedAttributes :: Lens' DescribeAutoMLJobResponse (Maybe ResolvedAttributes)

This contains ProblemType, AutoMLJobObjective, and CompletionCriteria. If you do not provide these values, they are auto-inferred. If you do provide them, the values used are the ones you provide.

describeAutoMLJobResponse_autoMLJobArtifacts :: Lens' DescribeAutoMLJobResponse (Maybe AutoMLJobArtifacts)

Returns information on the job's artifacts found in AutoMLJobArtifacts.

describeAutoMLJobResponse_partialFailureReasons :: Lens' DescribeAutoMLJobResponse (Maybe (NonEmpty AutoMLPartialFailureReason))

Returns a list of reasons for partial failures within an AutoML job.

describeAutoMLJobResponse_failureReason :: Lens' DescribeAutoMLJobResponse (Maybe Text)

Returns the failure reason for an AutoML job, when applicable.

describeAutoMLJobResponse_autoMLJobObjective :: Lens' DescribeAutoMLJobResponse (Maybe AutoMLJobObjective)

Returns the job's objective.

describeAutoMLJobResponse_modelDeployConfig :: Lens' DescribeAutoMLJobResponse (Maybe ModelDeployConfig)

Indicates whether the model was deployed automatically to an endpoint and the name of that endpoint if deployed automatically.

describeAutoMLJobResponse_autoMLJobConfig :: Lens' DescribeAutoMLJobResponse (Maybe AutoMLJobConfig)

Returns the configuration for the AutoML job.

describeAutoMLJobResponse_problemType :: Lens' DescribeAutoMLJobResponse (Maybe ProblemType)

Returns the job's problem type.

describeAutoMLJobResponse_bestCandidate :: Lens' DescribeAutoMLJobResponse (Maybe AutoMLCandidate)

Returns the job's best AutoMLCandidate.

describeAutoMLJobResponse_httpStatus :: Lens' DescribeAutoMLJobResponse Int

The response's http status code.

describeAutoMLJobResponse_autoMLJobName :: Lens' DescribeAutoMLJobResponse Text

Returns the name of the AutoML job.

describeAutoMLJobResponse_autoMLJobArn :: Lens' DescribeAutoMLJobResponse Text

Returns the ARN of the AutoML job.

describeAutoMLJobResponse_inputDataConfig :: Lens' DescribeAutoMLJobResponse (NonEmpty AutoMLChannel)

Returns the input data configuration for the AutoML job..

describeAutoMLJobResponse_outputDataConfig :: Lens' DescribeAutoMLJobResponse AutoMLOutputDataConfig

Returns the job's output data config.

describeAutoMLJobResponse_roleArn :: Lens' DescribeAutoMLJobResponse Text

The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that has read permission to the input data location and write permission to the output data location in Amazon S3.

describeAutoMLJobResponse_creationTime :: Lens' DescribeAutoMLJobResponse UTCTime

Returns the creation time of the AutoML job.

describeAutoMLJobResponse_lastModifiedTime :: Lens' DescribeAutoMLJobResponse UTCTime

Returns the job's last modified time.

describeAutoMLJobResponse_autoMLJobStatus :: Lens' DescribeAutoMLJobResponse AutoMLJobStatus

Returns the status of the AutoML job.

describeAutoMLJobResponse_autoMLJobSecondaryStatus :: Lens' DescribeAutoMLJobResponse AutoMLJobSecondaryStatus

Returns the secondary status of the AutoML job.