Network.AWS.SageMaker.DescribeCodeRepository

Description

Gets details about the specified Git repository.

Creating a Request

data DescribeCodeRepository

See: newDescribeCodeRepository smart constructor.

Constructors

DescribeCodeRepository'

Fields

codeRepositoryName :: Text

The name of the Git repository to describe.

newDescribeCodeRepository

Arguments

:: Text
$sel:codeRepositoryName:DescribeCodeRepository'

-> DescribeCodeRepository

Create a value of DescribeCodeRepository with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeRepositoryName:DescribeCodeRepository', describeCodeRepository_codeRepositoryName - The name of the Git repository to describe.

Request Lenses

describeCodeRepository_codeRepositoryName :: Lens' DescribeCodeRepository Text

The name of the Git repository to describe.

Destructuring the Response

data DescribeCodeRepositoryResponse

See: newDescribeCodeRepositoryResponse smart constructor.

Constructors

DescribeCodeRepositoryResponse'

Fields

gitConfig :: Maybe GitConfig

Configuration details about the repository, including the URL where the repository is located, the default branch, and the Amazon Resource Name (ARN) of the Amazon Web Services Secrets Manager secret that contains the credentials used to access the repository.

httpStatus :: Int

The response's http status code.

codeRepositoryName :: Text

The name of the Git repository.

codeRepositoryArn :: Text

The Amazon Resource Name (ARN) of the Git repository.

creationTime :: POSIX

The date and time that the repository was created.

lastModifiedTime :: POSIX

The date and time that the repository was last changed.

newDescribeCodeRepositoryResponse

Arguments

:: Int
$sel:httpStatus:DescribeCodeRepositoryResponse'

-> Text
$sel:codeRepositoryName:DescribeCodeRepository'

-> Text
$sel:codeRepositoryArn:DescribeCodeRepositoryResponse'

-> UTCTime
$sel:creationTime:DescribeCodeRepositoryResponse'

-> UTCTime
$sel:lastModifiedTime:DescribeCodeRepositoryResponse'

-> DescribeCodeRepositoryResponse

Create a value of DescribeCodeRepositoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gitConfig:DescribeCodeRepositoryResponse', describeCodeRepositoryResponse_gitConfig - Configuration details about the repository, including the URL where the repository is located, the default branch, and the Amazon Resource Name (ARN) of the Amazon Web Services Secrets Manager secret that contains the credentials used to access the repository.

$sel:httpStatus:DescribeCodeRepositoryResponse', describeCodeRepositoryResponse_httpStatus - The response's http status code.

$sel:codeRepositoryName:DescribeCodeRepository', describeCodeRepositoryResponse_codeRepositoryName - The name of the Git repository.

$sel:codeRepositoryArn:DescribeCodeRepositoryResponse', describeCodeRepositoryResponse_codeRepositoryArn - The Amazon Resource Name (ARN) of the Git repository.

$sel:creationTime:DescribeCodeRepositoryResponse', describeCodeRepositoryResponse_creationTime - The date and time that the repository was created.

$sel:lastModifiedTime:DescribeCodeRepositoryResponse', describeCodeRepositoryResponse_lastModifiedTime - The date and time that the repository was last changed.

Response Lenses

describeCodeRepositoryResponse_gitConfig :: Lens' DescribeCodeRepositoryResponse (Maybe GitConfig)

Configuration details about the repository, including the URL where the repository is located, the default branch, and the Amazon Resource Name (ARN) of the Amazon Web Services Secrets Manager secret that contains the credentials used to access the repository.

describeCodeRepositoryResponse_httpStatus :: Lens' DescribeCodeRepositoryResponse Int

The response's http status code.

describeCodeRepositoryResponse_codeRepositoryName :: Lens' DescribeCodeRepositoryResponse Text

The name of the Git repository.

describeCodeRepositoryResponse_codeRepositoryArn :: Lens' DescribeCodeRepositoryResponse Text

The Amazon Resource Name (ARN) of the Git repository.

describeCodeRepositoryResponse_creationTime :: Lens' DescribeCodeRepositoryResponse UTCTime

The date and time that the repository was created.

describeCodeRepositoryResponse_lastModifiedTime :: Lens' DescribeCodeRepositoryResponse UTCTime

The date and time that the repository was last changed.