Network.AWS.SageMaker.DescribeDataQualityJobDefinition

Description

Gets the details of a data quality monitoring job definition.

Creating a Request

data DescribeDataQualityJobDefinition

See: newDescribeDataQualityJobDefinition smart constructor.

Constructors

DescribeDataQualityJobDefinition'

Fields

jobDefinitionName :: Text

The name of the data quality monitoring job definition to describe.

newDescribeDataQualityJobDefinition

Arguments

:: Text
$sel:jobDefinitionName:DescribeDataQualityJobDefinition'

-> DescribeDataQualityJobDefinition

Create a value of DescribeDataQualityJobDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobDefinitionName:DescribeDataQualityJobDefinition', describeDataQualityJobDefinition_jobDefinitionName - The name of the data quality monitoring job definition to describe.

Request Lenses

describeDataQualityJobDefinition_jobDefinitionName :: Lens' DescribeDataQualityJobDefinition Text

The name of the data quality monitoring job definition to describe.

Destructuring the Response

data DescribeDataQualityJobDefinitionResponse

See: newDescribeDataQualityJobDefinitionResponse smart constructor.

Constructors

DescribeDataQualityJobDefinitionResponse'

Fields

networkConfig :: Maybe MonitoringNetworkConfig

The networking configuration for the data quality monitoring job.

dataQualityBaselineConfig :: Maybe DataQualityBaselineConfig

The constraints and baselines for the data quality monitoring job definition.

stoppingCondition :: Maybe MonitoringStoppingCondition

httpStatus :: Int

The response's http status code.

jobDefinitionArn :: Text

The Amazon Resource Name (ARN) of the data quality monitoring job definition.

jobDefinitionName :: Text

The name of the data quality monitoring job definition.

creationTime :: POSIX

The time that the data quality monitoring job definition was created.

dataQualityAppSpecification :: DataQualityAppSpecification

Information about the container that runs the data quality monitoring job.

dataQualityJobInput :: DataQualityJobInput

The list of inputs for the data quality monitoring job. Currently endpoints are supported.

dataQualityJobOutputConfig :: MonitoringOutputConfig

jobResources :: MonitoringResources

roleArn :: Text

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

newDescribeDataQualityJobDefinitionResponse

Arguments

:: Int
$sel:httpStatus:DescribeDataQualityJobDefinitionResponse'

-> Text
$sel:jobDefinitionArn:DescribeDataQualityJobDefinitionResponse'

-> Text
$sel:jobDefinitionName:DescribeDataQualityJobDefinition'

-> UTCTime
$sel:creationTime:DescribeDataQualityJobDefinitionResponse'

-> DataQualityAppSpecification
$sel:dataQualityAppSpecification:DescribeDataQualityJobDefinitionResponse'

-> DataQualityJobInput
$sel:dataQualityJobInput:DescribeDataQualityJobDefinitionResponse'

-> MonitoringOutputConfig
$sel:dataQualityJobOutputConfig:DescribeDataQualityJobDefinitionResponse'

-> MonitoringResources
$sel:jobResources:DescribeDataQualityJobDefinitionResponse'

-> Text
$sel:roleArn:DescribeDataQualityJobDefinitionResponse'

-> DescribeDataQualityJobDefinitionResponse

Create a value of DescribeDataQualityJobDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkConfig:DescribeDataQualityJobDefinitionResponse', describeDataQualityJobDefinitionResponse_networkConfig - The networking configuration for the data quality monitoring job.

$sel:dataQualityBaselineConfig:DescribeDataQualityJobDefinitionResponse', describeDataQualityJobDefinitionResponse_dataQualityBaselineConfig - The constraints and baselines for the data quality monitoring job definition.

$sel:stoppingCondition:DescribeDataQualityJobDefinitionResponse', describeDataQualityJobDefinitionResponse_stoppingCondition - Undocumented member.

$sel:httpStatus:DescribeDataQualityJobDefinitionResponse', describeDataQualityJobDefinitionResponse_httpStatus - The response's http status code.

$sel:jobDefinitionArn:DescribeDataQualityJobDefinitionResponse', describeDataQualityJobDefinitionResponse_jobDefinitionArn - The Amazon Resource Name (ARN) of the data quality monitoring job definition.

$sel:jobDefinitionName:DescribeDataQualityJobDefinition', describeDataQualityJobDefinitionResponse_jobDefinitionName - The name of the data quality monitoring job definition.

$sel:creationTime:DescribeDataQualityJobDefinitionResponse', describeDataQualityJobDefinitionResponse_creationTime - The time that the data quality monitoring job definition was created.

$sel:dataQualityAppSpecification:DescribeDataQualityJobDefinitionResponse', describeDataQualityJobDefinitionResponse_dataQualityAppSpecification - Information about the container that runs the data quality monitoring job.

$sel:dataQualityJobInput:DescribeDataQualityJobDefinitionResponse', describeDataQualityJobDefinitionResponse_dataQualityJobInput - The list of inputs for the data quality monitoring job. Currently endpoints are supported.

$sel:dataQualityJobOutputConfig:DescribeDataQualityJobDefinitionResponse', describeDataQualityJobDefinitionResponse_dataQualityJobOutputConfig - Undocumented member.

$sel:jobResources:DescribeDataQualityJobDefinitionResponse', describeDataQualityJobDefinitionResponse_jobResources - Undocumented member.

$sel:roleArn:DescribeDataQualityJobDefinitionResponse', describeDataQualityJobDefinitionResponse_roleArn - The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

Response Lenses

describeDataQualityJobDefinitionResponse_networkConfig :: Lens' DescribeDataQualityJobDefinitionResponse (Maybe MonitoringNetworkConfig)

The networking configuration for the data quality monitoring job.

describeDataQualityJobDefinitionResponse_dataQualityBaselineConfig :: Lens' DescribeDataQualityJobDefinitionResponse (Maybe DataQualityBaselineConfig)

The constraints and baselines for the data quality monitoring job definition.

describeDataQualityJobDefinitionResponse_stoppingCondition :: Lens' DescribeDataQualityJobDefinitionResponse (Maybe MonitoringStoppingCondition)

Undocumented member.

describeDataQualityJobDefinitionResponse_httpStatus :: Lens' DescribeDataQualityJobDefinitionResponse Int

The response's http status code.

describeDataQualityJobDefinitionResponse_jobDefinitionArn :: Lens' DescribeDataQualityJobDefinitionResponse Text

The Amazon Resource Name (ARN) of the data quality monitoring job definition.

describeDataQualityJobDefinitionResponse_jobDefinitionName :: Lens' DescribeDataQualityJobDefinitionResponse Text

The name of the data quality monitoring job definition.

describeDataQualityJobDefinitionResponse_creationTime :: Lens' DescribeDataQualityJobDefinitionResponse UTCTime

The time that the data quality monitoring job definition was created.

describeDataQualityJobDefinitionResponse_dataQualityAppSpecification :: Lens' DescribeDataQualityJobDefinitionResponse DataQualityAppSpecification

Information about the container that runs the data quality monitoring job.

describeDataQualityJobDefinitionResponse_dataQualityJobInput :: Lens' DescribeDataQualityJobDefinitionResponse DataQualityJobInput

The list of inputs for the data quality monitoring job. Currently endpoints are supported.

describeDataQualityJobDefinitionResponse_dataQualityJobOutputConfig :: Lens' DescribeDataQualityJobDefinitionResponse MonitoringOutputConfig

Undocumented member.

describeDataQualityJobDefinitionResponse_jobResources :: Lens' DescribeDataQualityJobDefinitionResponse MonitoringResources

Undocumented member.

describeDataQualityJobDefinitionResponse_roleArn :: Lens' DescribeDataQualityJobDefinitionResponse Text

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.