Network.AWS.SageMaker.DescribeDevice

Description

Describes the device.

Creating a Request

data DescribeDevice

See: newDescribeDevice smart constructor.

Constructors

DescribeDevice'

Fields

nextToken :: Maybe Text

Next token of device description.

deviceName :: Text

The unique ID of the device.

deviceFleetName :: Text

The name of the fleet the devices belong to.

newDescribeDevice

Arguments

:: Text
$sel:deviceName:DescribeDevice'

-> Text
$sel:deviceFleetName:DescribeDevice'

-> DescribeDevice

Create a value of DescribeDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeDevice', describeDevice_nextToken - Next token of device description.

$sel:deviceName:DescribeDevice', describeDevice_deviceName - The unique ID of the device.

$sel:deviceFleetName:DescribeDevice', describeDevice_deviceFleetName - The name of the fleet the devices belong to.

Request Lenses

describeDevice_nextToken :: Lens' DescribeDevice (Maybe Text)

Next token of device description.

describeDevice_deviceName :: Lens' DescribeDevice Text

The unique ID of the device.

describeDevice_deviceFleetName :: Lens' DescribeDevice Text

The name of the fleet the devices belong to.

Destructuring the Response

data DescribeDeviceResponse

See: newDescribeDeviceResponse smart constructor.

Constructors

DescribeDeviceResponse'

Fields

deviceArn :: Maybe Text

The Amazon Resource Name (ARN) of the device.

nextToken :: Maybe Text

The response from the last list when returning a list large enough to need tokening.

models :: Maybe [EdgeModel]

Models on the device.

maxModels :: Maybe Int

The maximum number of models.

latestHeartbeat :: Maybe POSIX

The last heartbeat received from the device.

description :: Maybe Text

A description of the device.

iotThingName :: Maybe Text

The Amazon Web Services Internet of Things (IoT) object thing name associated with the device.

httpStatus :: Int

The response's http status code.

deviceName :: Text

The unique identifier of the device.

deviceFleetName :: Text

The name of the fleet the device belongs to.

registrationTime :: POSIX

The timestamp of the last registration or de-reregistration.

newDescribeDeviceResponse

Arguments

:: Int
$sel:httpStatus:DescribeDeviceResponse'

-> Text
$sel:deviceName:DescribeDevice'

-> Text
$sel:deviceFleetName:DescribeDevice'

-> UTCTime
$sel:registrationTime:DescribeDeviceResponse'

-> DescribeDeviceResponse

Create a value of DescribeDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceArn:DescribeDeviceResponse', describeDeviceResponse_deviceArn - The Amazon Resource Name (ARN) of the device.

$sel:nextToken:DescribeDevice', describeDeviceResponse_nextToken - The response from the last list when returning a list large enough to need tokening.

$sel:models:DescribeDeviceResponse', describeDeviceResponse_models - Models on the device.

$sel:maxModels:DescribeDeviceResponse', describeDeviceResponse_maxModels - The maximum number of models.

$sel:latestHeartbeat:DescribeDeviceResponse', describeDeviceResponse_latestHeartbeat - The last heartbeat received from the device.

$sel:description:DescribeDeviceResponse', describeDeviceResponse_description - A description of the device.

$sel:iotThingName:DescribeDeviceResponse', describeDeviceResponse_iotThingName - The Amazon Web Services Internet of Things (IoT) object thing name associated with the device.

$sel:httpStatus:DescribeDeviceResponse', describeDeviceResponse_httpStatus - The response's http status code.

$sel:deviceName:DescribeDevice', describeDeviceResponse_deviceName - The unique identifier of the device.

$sel:deviceFleetName:DescribeDevice', describeDeviceResponse_deviceFleetName - The name of the fleet the device belongs to.

$sel:registrationTime:DescribeDeviceResponse', describeDeviceResponse_registrationTime - The timestamp of the last registration or de-reregistration.

Response Lenses

describeDeviceResponse_deviceArn :: Lens' DescribeDeviceResponse (Maybe Text)

The Amazon Resource Name (ARN) of the device.

describeDeviceResponse_nextToken :: Lens' DescribeDeviceResponse (Maybe Text)

The response from the last list when returning a list large enough to need tokening.

describeDeviceResponse_models :: Lens' DescribeDeviceResponse (Maybe [EdgeModel])

Models on the device.

describeDeviceResponse_maxModels :: Lens' DescribeDeviceResponse (Maybe Int)

The maximum number of models.

describeDeviceResponse_latestHeartbeat :: Lens' DescribeDeviceResponse (Maybe UTCTime)

The last heartbeat received from the device.

describeDeviceResponse_description :: Lens' DescribeDeviceResponse (Maybe Text)

A description of the device.

describeDeviceResponse_iotThingName :: Lens' DescribeDeviceResponse (Maybe Text)

The Amazon Web Services Internet of Things (IoT) object thing name associated with the device.

describeDeviceResponse_httpStatus :: Lens' DescribeDeviceResponse Int

The response's http status code.

describeDeviceResponse_deviceName :: Lens' DescribeDeviceResponse Text

The unique identifier of the device.

describeDeviceResponse_deviceFleetName :: Lens' DescribeDeviceResponse Text

The name of the fleet the device belongs to.

describeDeviceResponse_registrationTime :: Lens' DescribeDeviceResponse UTCTime

The timestamp of the last registration or de-reregistration.