Network.AWS.SageMaker.DescribeDomain

Description

The description of the domain.

Creating a Request

data DescribeDomain

See: newDescribeDomain smart constructor.

Constructors

DescribeDomain'

Fields

domainId :: Text

The domain ID.

newDescribeDomain

Arguments

:: Text
$sel:domainId:DescribeDomain'

-> DescribeDomain

Create a value of DescribeDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainId:DescribeDomain', describeDomain_domainId - The domain ID.

Request Lenses

describeDomain_domainId :: Lens' DescribeDomain Text

The domain ID.

Destructuring the Response

data DescribeDomainResponse

See: newDescribeDomainResponse smart constructor.

Constructors

DescribeDomainResponse'

Fields

creationTime :: Maybe POSIX

The creation time.

status :: Maybe DomainStatus

The status.

singleSignOnManagedApplicationInstanceId :: Maybe Text

The SSO managed application instance ID.

authMode :: Maybe AuthMode

The domain's authentication mode.

domainArn :: Maybe Text

The domain's Amazon Resource Name (ARN).

subnetIds :: Maybe (NonEmpty Text)

The VPC subnets that Studio uses for communication.

domainId :: Maybe Text

The domain ID.

kmsKeyId :: Maybe Text

The Amazon Web Services KMS customer managed key used to encrypt the EFS volume attached to the domain.

domainName :: Maybe Text

The domain name.

defaultUserSettings :: Maybe UserSettings

Settings which are applied to UserProfiles in this domain if settings are not explicitly specified in a given UserProfile.

failureReason :: Maybe Text

The failure reason.

homeEfsFileSystemId :: Maybe Text

The ID of the Amazon Elastic File System (EFS) managed by this Domain.

lastModifiedTime :: Maybe POSIX

The last modified time.

url :: Maybe Text

The domain's URL.

homeEfsFileSystemKmsKeyId :: Maybe Text

This member is deprecated and replaced with KmsKeyId.

appNetworkAccessType :: Maybe AppNetworkAccessType

Specifies the VPC used for non-EFS traffic. The default value is PublicInternetOnly.

PublicInternetOnly - Non-EFS traffic is through a VPC managed by Amazon SageMaker, which allows direct internet access

VpcOnly - All Studio traffic is through the specified VPC and subnets

vpcId :: Maybe Text

The ID of the Amazon Virtual Private Cloud (VPC) that Studio uses for communication.

httpStatus :: Int

The response's http status code.

newDescribeDomainResponse

Arguments

:: Int
$sel:httpStatus:DescribeDomainResponse'

-> DescribeDomainResponse

Create a value of DescribeDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeDomainResponse', describeDomainResponse_creationTime - The creation time.

$sel:status:DescribeDomainResponse', describeDomainResponse_status - The status.

$sel:singleSignOnManagedApplicationInstanceId:DescribeDomainResponse', describeDomainResponse_singleSignOnManagedApplicationInstanceId - The SSO managed application instance ID.

$sel:authMode:DescribeDomainResponse', describeDomainResponse_authMode - The domain's authentication mode.

$sel:domainArn:DescribeDomainResponse', describeDomainResponse_domainArn - The domain's Amazon Resource Name (ARN).

$sel:subnetIds:DescribeDomainResponse', describeDomainResponse_subnetIds - The VPC subnets that Studio uses for communication.

$sel:domainId:DescribeDomain', describeDomainResponse_domainId - The domain ID.

$sel:kmsKeyId:DescribeDomainResponse', describeDomainResponse_kmsKeyId - The Amazon Web Services KMS customer managed key used to encrypt the EFS volume attached to the domain.

$sel:domainName:DescribeDomainResponse', describeDomainResponse_domainName - The domain name.

$sel:defaultUserSettings:DescribeDomainResponse', describeDomainResponse_defaultUserSettings - Settings which are applied to UserProfiles in this domain if settings are not explicitly specified in a given UserProfile.

$sel:failureReason:DescribeDomainResponse', describeDomainResponse_failureReason - The failure reason.

$sel:homeEfsFileSystemId:DescribeDomainResponse', describeDomainResponse_homeEfsFileSystemId - The ID of the Amazon Elastic File System (EFS) managed by this Domain.

$sel:lastModifiedTime:DescribeDomainResponse', describeDomainResponse_lastModifiedTime - The last modified time.

$sel:url:DescribeDomainResponse', describeDomainResponse_url - The domain's URL.

$sel:homeEfsFileSystemKmsKeyId:DescribeDomainResponse', describeDomainResponse_homeEfsFileSystemKmsKeyId - This member is deprecated and replaced with KmsKeyId.

$sel:appNetworkAccessType:DescribeDomainResponse', describeDomainResponse_appNetworkAccessType - Specifies the VPC used for non-EFS traffic. The default value is PublicInternetOnly.

PublicInternetOnly - Non-EFS traffic is through a VPC managed by Amazon SageMaker, which allows direct internet access

VpcOnly - All Studio traffic is through the specified VPC and subnets

$sel:vpcId:DescribeDomainResponse', describeDomainResponse_vpcId - The ID of the Amazon Virtual Private Cloud (VPC) that Studio uses for communication.

$sel:httpStatus:DescribeDomainResponse', describeDomainResponse_httpStatus - The response's http status code.

Response Lenses

describeDomainResponse_creationTime :: Lens' DescribeDomainResponse (Maybe UTCTime)

The creation time.

describeDomainResponse_status :: Lens' DescribeDomainResponse (Maybe DomainStatus)

The status.

describeDomainResponse_singleSignOnManagedApplicationInstanceId :: Lens' DescribeDomainResponse (Maybe Text)

The SSO managed application instance ID.

describeDomainResponse_authMode :: Lens' DescribeDomainResponse (Maybe AuthMode)

The domain's authentication mode.

describeDomainResponse_domainArn :: Lens' DescribeDomainResponse (Maybe Text)

The domain's Amazon Resource Name (ARN).

describeDomainResponse_subnetIds :: Lens' DescribeDomainResponse (Maybe (NonEmpty Text))

The VPC subnets that Studio uses for communication.

describeDomainResponse_domainId :: Lens' DescribeDomainResponse (Maybe Text)

The domain ID.

describeDomainResponse_kmsKeyId :: Lens' DescribeDomainResponse (Maybe Text)

The Amazon Web Services KMS customer managed key used to encrypt the EFS volume attached to the domain.

describeDomainResponse_domainName :: Lens' DescribeDomainResponse (Maybe Text)

The domain name.

describeDomainResponse_defaultUserSettings :: Lens' DescribeDomainResponse (Maybe UserSettings)

Settings which are applied to UserProfiles in this domain if settings are not explicitly specified in a given UserProfile.

describeDomainResponse_failureReason :: Lens' DescribeDomainResponse (Maybe Text)

The failure reason.

describeDomainResponse_homeEfsFileSystemId :: Lens' DescribeDomainResponse (Maybe Text)

The ID of the Amazon Elastic File System (EFS) managed by this Domain.

describeDomainResponse_lastModifiedTime :: Lens' DescribeDomainResponse (Maybe UTCTime)

The last modified time.

describeDomainResponse_url :: Lens' DescribeDomainResponse (Maybe Text)

The domain's URL.

describeDomainResponse_homeEfsFileSystemKmsKeyId :: Lens' DescribeDomainResponse (Maybe Text)

This member is deprecated and replaced with KmsKeyId.

describeDomainResponse_appNetworkAccessType :: Lens' DescribeDomainResponse (Maybe AppNetworkAccessType)

Specifies the VPC used for non-EFS traffic. The default value is PublicInternetOnly.

PublicInternetOnly - Non-EFS traffic is through a VPC managed by Amazon SageMaker, which allows direct internet access

VpcOnly - All Studio traffic is through the specified VPC and subnets

describeDomainResponse_vpcId :: Lens' DescribeDomainResponse (Maybe Text)

The ID of the Amazon Virtual Private Cloud (VPC) that Studio uses for communication.

describeDomainResponse_httpStatus :: Lens' DescribeDomainResponse Int

The response's http status code.