Network.AWS.SageMaker.DescribeEdgePackagingJob

Description

A description of edge packaging jobs.

Creating a Request

data DescribeEdgePackagingJob

See: newDescribeEdgePackagingJob smart constructor.

Constructors

DescribeEdgePackagingJob'

Fields

edgePackagingJobName :: Text

The name of the edge packaging job.

newDescribeEdgePackagingJob

Arguments

:: Text
$sel:edgePackagingJobName:DescribeEdgePackagingJob'

-> DescribeEdgePackagingJob

Create a value of DescribeEdgePackagingJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:edgePackagingJobName:DescribeEdgePackagingJob', describeEdgePackagingJob_edgePackagingJobName - The name of the edge packaging job.

Request Lenses

describeEdgePackagingJob_edgePackagingJobName :: Lens' DescribeEdgePackagingJob Text

The name of the edge packaging job.

Destructuring the Response

data DescribeEdgePackagingJobResponse

See: newDescribeEdgePackagingJobResponse smart constructor.

Constructors

DescribeEdgePackagingJobResponse'

Fields

creationTime :: Maybe POSIX

The timestamp of when the packaging job was created.

resourceKey :: Maybe Text

The Amazon Web Services KMS key to use when encrypting the EBS volume the job run on.

modelName :: Maybe Text

The name of the model.

modelSignature :: Maybe Text

The signature document of files in the model artifact.

lastModifiedTime :: Maybe POSIX

The timestamp of when the job was last updated.

modelArtifact :: Maybe Text

The Amazon Simple Storage (S3) URI where model artifacts ares stored.

compilationJobName :: Maybe Text

The name of the SageMaker Neo compilation job that is used to locate model artifacts that are being packaged.

edgePackagingJobStatusMessage :: Maybe Text

Returns a message describing the job status and error messages.

outputConfig :: Maybe EdgeOutputConfig

The output configuration for the edge packaging job.

modelVersion :: Maybe Text

The version of the model.

presetDeploymentOutput :: Maybe EdgePresetDeploymentOutput

The output of a SageMaker Edge Manager deployable resource.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of an IAM role that enables Amazon SageMaker to download and upload the model, and to contact Neo.

httpStatus :: Int

The response's http status code.

edgePackagingJobArn :: Text

The Amazon Resource Name (ARN) of the edge packaging job.

edgePackagingJobName :: Text

The name of the edge packaging job.

edgePackagingJobStatus :: EdgePackagingJobStatus

The current status of the packaging job.

newDescribeEdgePackagingJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeEdgePackagingJobResponse'

-> Text
$sel:edgePackagingJobArn:DescribeEdgePackagingJobResponse'

-> Text
$sel:edgePackagingJobName:DescribeEdgePackagingJob'

-> EdgePackagingJobStatus
$sel:edgePackagingJobStatus:DescribeEdgePackagingJobResponse'

-> DescribeEdgePackagingJobResponse

Create a value of DescribeEdgePackagingJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeEdgePackagingJobResponse', describeEdgePackagingJobResponse_creationTime - The timestamp of when the packaging job was created.

$sel:resourceKey:DescribeEdgePackagingJobResponse', describeEdgePackagingJobResponse_resourceKey - The Amazon Web Services KMS key to use when encrypting the EBS volume the job run on.

$sel:modelName:DescribeEdgePackagingJobResponse', describeEdgePackagingJobResponse_modelName - The name of the model.

$sel:modelSignature:DescribeEdgePackagingJobResponse', describeEdgePackagingJobResponse_modelSignature - The signature document of files in the model artifact.

$sel:lastModifiedTime:DescribeEdgePackagingJobResponse', describeEdgePackagingJobResponse_lastModifiedTime - The timestamp of when the job was last updated.

$sel:modelArtifact:DescribeEdgePackagingJobResponse', describeEdgePackagingJobResponse_modelArtifact - The Amazon Simple Storage (S3) URI where model artifacts ares stored.

$sel:compilationJobName:DescribeEdgePackagingJobResponse', describeEdgePackagingJobResponse_compilationJobName - The name of the SageMaker Neo compilation job that is used to locate model artifacts that are being packaged.

$sel:edgePackagingJobStatusMessage:DescribeEdgePackagingJobResponse', describeEdgePackagingJobResponse_edgePackagingJobStatusMessage - Returns a message describing the job status and error messages.

$sel:outputConfig:DescribeEdgePackagingJobResponse', describeEdgePackagingJobResponse_outputConfig - The output configuration for the edge packaging job.

$sel:modelVersion:DescribeEdgePackagingJobResponse', describeEdgePackagingJobResponse_modelVersion - The version of the model.

$sel:presetDeploymentOutput:DescribeEdgePackagingJobResponse', describeEdgePackagingJobResponse_presetDeploymentOutput - The output of a SageMaker Edge Manager deployable resource.

$sel:roleArn:DescribeEdgePackagingJobResponse', describeEdgePackagingJobResponse_roleArn - The Amazon Resource Name (ARN) of an IAM role that enables Amazon SageMaker to download and upload the model, and to contact Neo.

$sel:httpStatus:DescribeEdgePackagingJobResponse', describeEdgePackagingJobResponse_httpStatus - The response's http status code.

$sel:edgePackagingJobArn:DescribeEdgePackagingJobResponse', describeEdgePackagingJobResponse_edgePackagingJobArn - The Amazon Resource Name (ARN) of the edge packaging job.

$sel:edgePackagingJobName:DescribeEdgePackagingJob', describeEdgePackagingJobResponse_edgePackagingJobName - The name of the edge packaging job.

$sel:edgePackagingJobStatus:DescribeEdgePackagingJobResponse', describeEdgePackagingJobResponse_edgePackagingJobStatus - The current status of the packaging job.

Response Lenses

describeEdgePackagingJobResponse_creationTime :: Lens' DescribeEdgePackagingJobResponse (Maybe UTCTime)

The timestamp of when the packaging job was created.

describeEdgePackagingJobResponse_resourceKey :: Lens' DescribeEdgePackagingJobResponse (Maybe Text)

The Amazon Web Services KMS key to use when encrypting the EBS volume the job run on.

describeEdgePackagingJobResponse_modelName :: Lens' DescribeEdgePackagingJobResponse (Maybe Text)

The name of the model.

describeEdgePackagingJobResponse_modelSignature :: Lens' DescribeEdgePackagingJobResponse (Maybe Text)

The signature document of files in the model artifact.

describeEdgePackagingJobResponse_lastModifiedTime :: Lens' DescribeEdgePackagingJobResponse (Maybe UTCTime)

The timestamp of when the job was last updated.

describeEdgePackagingJobResponse_modelArtifact :: Lens' DescribeEdgePackagingJobResponse (Maybe Text)

The Amazon Simple Storage (S3) URI where model artifacts ares stored.

describeEdgePackagingJobResponse_compilationJobName :: Lens' DescribeEdgePackagingJobResponse (Maybe Text)

The name of the SageMaker Neo compilation job that is used to locate model artifacts that are being packaged.

describeEdgePackagingJobResponse_edgePackagingJobStatusMessage :: Lens' DescribeEdgePackagingJobResponse (Maybe Text)

Returns a message describing the job status and error messages.

describeEdgePackagingJobResponse_outputConfig :: Lens' DescribeEdgePackagingJobResponse (Maybe EdgeOutputConfig)

The output configuration for the edge packaging job.

describeEdgePackagingJobResponse_modelVersion :: Lens' DescribeEdgePackagingJobResponse (Maybe Text)

The version of the model.

describeEdgePackagingJobResponse_presetDeploymentOutput :: Lens' DescribeEdgePackagingJobResponse (Maybe EdgePresetDeploymentOutput)

The output of a SageMaker Edge Manager deployable resource.

describeEdgePackagingJobResponse_roleArn :: Lens' DescribeEdgePackagingJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of an IAM role that enables Amazon SageMaker to download and upload the model, and to contact Neo.

describeEdgePackagingJobResponse_httpStatus :: Lens' DescribeEdgePackagingJobResponse Int

The response's http status code.

describeEdgePackagingJobResponse_edgePackagingJobArn :: Lens' DescribeEdgePackagingJobResponse Text

The Amazon Resource Name (ARN) of the edge packaging job.

describeEdgePackagingJobResponse_edgePackagingJobName :: Lens' DescribeEdgePackagingJobResponse Text

The name of the edge packaging job.

describeEdgePackagingJobResponse_edgePackagingJobStatus :: Lens' DescribeEdgePackagingJobResponse EdgePackagingJobStatus

The current status of the packaging job.