Network.AWS.SageMaker.DescribeEndpoint

Description

Returns the description of an endpoint.

Creating a Request

data DescribeEndpoint

See: newDescribeEndpoint smart constructor.

Constructors

DescribeEndpoint'

Fields

endpointName :: Text

The name of the endpoint.

newDescribeEndpoint

Arguments

:: Text
$sel:endpointName:DescribeEndpoint'

-> DescribeEndpoint

Create a value of DescribeEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpointName:DescribeEndpoint', describeEndpoint_endpointName - The name of the endpoint.

Request Lenses

describeEndpoint_endpointName :: Lens' DescribeEndpoint Text

The name of the endpoint.

Destructuring the Response

data DescribeEndpointResponse

See: newDescribeEndpointResponse smart constructor.

Constructors

DescribeEndpointResponse'

Fields

asyncInferenceConfig :: Maybe AsyncInferenceConfig

Returns the description of an endpoint configuration created using the CreateEndpointConfig API.

productionVariants :: Maybe (NonEmpty ProductionVariantSummary)

An array of ProductionVariantSummary objects, one for each model hosted behind this endpoint.

lastDeploymentConfig :: Maybe DeploymentConfig

The most recent deployment configuration for the endpoint.

failureReason :: Maybe Text

If the status of the endpoint is Failed, the reason why it failed.

dataCaptureConfig :: Maybe DataCaptureConfigSummary

httpStatus :: Int

The response's http status code.

endpointName :: Text

Name of the endpoint.

endpointArn :: Text

The Amazon Resource Name (ARN) of the endpoint.

endpointConfigName :: Text

The name of the endpoint configuration associated with this endpoint.

endpointStatus :: EndpointStatus

The status of the endpoint.

OutOfService: Endpoint is not available to take incoming requests.

Creating: CreateEndpoint is executing.

Updating: UpdateEndpoint or UpdateEndpointWeightsAndCapacities is executing.

SystemUpdating: Endpoint is undergoing maintenance and cannot be updated or deleted or re-scaled until it has completed. This maintenance operation does not change any customer-specified values such as VPC config, KMS encryption, model, instance type, or instance count.

RollingBack: Endpoint fails to scale up or down or change its variant weight and is in the process of rolling back to its previous configuration. Once the rollback completes, endpoint returns to an InService status. This transitional status only applies to an endpoint that has autoscaling enabled and is undergoing variant weight or capacity changes as part of an UpdateEndpointWeightsAndCapacities call or when the UpdateEndpointWeightsAndCapacities operation is called explicitly.

InService: Endpoint is available to process incoming requests.

Deleting: DeleteEndpoint is executing.

Failed: Endpoint could not be created, updated, or re-scaled. Use DescribeEndpointOutput$FailureReason for information about the failure. DeleteEndpoint is the only operation that can be performed on a failed endpoint.

creationTime :: POSIX

A timestamp that shows when the endpoint was created.

lastModifiedTime :: POSIX

A timestamp that shows when the endpoint was last modified.

newDescribeEndpointResponse

Arguments

:: Int
$sel:httpStatus:DescribeEndpointResponse'

-> Text
$sel:endpointName:DescribeEndpoint'

-> Text
$sel:endpointArn:DescribeEndpointResponse'

-> Text
$sel:endpointConfigName:DescribeEndpointResponse'

-> EndpointStatus
$sel:endpointStatus:DescribeEndpointResponse'

-> UTCTime
$sel:creationTime:DescribeEndpointResponse'

-> UTCTime
$sel:lastModifiedTime:DescribeEndpointResponse'

-> DescribeEndpointResponse

Create a value of DescribeEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:asyncInferenceConfig:DescribeEndpointResponse', describeEndpointResponse_asyncInferenceConfig - Returns the description of an endpoint configuration created using the CreateEndpointConfig API.

$sel:productionVariants:DescribeEndpointResponse', describeEndpointResponse_productionVariants - An array of ProductionVariantSummary objects, one for each model hosted behind this endpoint.

$sel:lastDeploymentConfig:DescribeEndpointResponse', describeEndpointResponse_lastDeploymentConfig - The most recent deployment configuration for the endpoint.

$sel:failureReason:DescribeEndpointResponse', describeEndpointResponse_failureReason - If the status of the endpoint is Failed, the reason why it failed.

$sel:dataCaptureConfig:DescribeEndpointResponse', describeEndpointResponse_dataCaptureConfig - Undocumented member.

$sel:httpStatus:DescribeEndpointResponse', describeEndpointResponse_httpStatus - The response's http status code.

$sel:endpointName:DescribeEndpoint', describeEndpointResponse_endpointName - Name of the endpoint.

$sel:endpointArn:DescribeEndpointResponse', describeEndpointResponse_endpointArn - The Amazon Resource Name (ARN) of the endpoint.

$sel:endpointConfigName:DescribeEndpointResponse', describeEndpointResponse_endpointConfigName - The name of the endpoint configuration associated with this endpoint.

$sel:endpointStatus:DescribeEndpointResponse', describeEndpointResponse_endpointStatus - The status of the endpoint.

OutOfService: Endpoint is not available to take incoming requests.

Creating: CreateEndpoint is executing.

Updating: UpdateEndpoint or UpdateEndpointWeightsAndCapacities is executing.

SystemUpdating: Endpoint is undergoing maintenance and cannot be updated or deleted or re-scaled until it has completed. This maintenance operation does not change any customer-specified values such as VPC config, KMS encryption, model, instance type, or instance count.

RollingBack: Endpoint fails to scale up or down or change its variant weight and is in the process of rolling back to its previous configuration. Once the rollback completes, endpoint returns to an InService status. This transitional status only applies to an endpoint that has autoscaling enabled and is undergoing variant weight or capacity changes as part of an UpdateEndpointWeightsAndCapacities call or when the UpdateEndpointWeightsAndCapacities operation is called explicitly.

InService: Endpoint is available to process incoming requests.

Deleting: DeleteEndpoint is executing.

Failed: Endpoint could not be created, updated, or re-scaled. Use DescribeEndpointOutput$FailureReason for information about the failure. DeleteEndpoint is the only operation that can be performed on a failed endpoint.

$sel:creationTime:DescribeEndpointResponse', describeEndpointResponse_creationTime - A timestamp that shows when the endpoint was created.

$sel:lastModifiedTime:DescribeEndpointResponse', describeEndpointResponse_lastModifiedTime - A timestamp that shows when the endpoint was last modified.

Response Lenses

describeEndpointResponse_asyncInferenceConfig :: Lens' DescribeEndpointResponse (Maybe AsyncInferenceConfig)

Returns the description of an endpoint configuration created using the CreateEndpointConfig API.

describeEndpointResponse_productionVariants :: Lens' DescribeEndpointResponse (Maybe (NonEmpty ProductionVariantSummary))

An array of ProductionVariantSummary objects, one for each model hosted behind this endpoint.

describeEndpointResponse_lastDeploymentConfig :: Lens' DescribeEndpointResponse (Maybe DeploymentConfig)

The most recent deployment configuration for the endpoint.

describeEndpointResponse_failureReason :: Lens' DescribeEndpointResponse (Maybe Text)

If the status of the endpoint is Failed, the reason why it failed.

describeEndpointResponse_dataCaptureConfig :: Lens' DescribeEndpointResponse (Maybe DataCaptureConfigSummary)

Undocumented member.

describeEndpointResponse_httpStatus :: Lens' DescribeEndpointResponse Int

The response's http status code.

describeEndpointResponse_endpointName :: Lens' DescribeEndpointResponse Text

Name of the endpoint.

describeEndpointResponse_endpointArn :: Lens' DescribeEndpointResponse Text

The Amazon Resource Name (ARN) of the endpoint.

describeEndpointResponse_endpointConfigName :: Lens' DescribeEndpointResponse Text

The name of the endpoint configuration associated with this endpoint.

describeEndpointResponse_endpointStatus :: Lens' DescribeEndpointResponse EndpointStatus

The status of the endpoint.

OutOfService: Endpoint is not available to take incoming requests.

Creating: CreateEndpoint is executing.

Updating: UpdateEndpoint or UpdateEndpointWeightsAndCapacities is executing.

SystemUpdating: Endpoint is undergoing maintenance and cannot be updated or deleted or re-scaled until it has completed. This maintenance operation does not change any customer-specified values such as VPC config, KMS encryption, model, instance type, or instance count.

RollingBack: Endpoint fails to scale up or down or change its variant weight and is in the process of rolling back to its previous configuration. Once the rollback completes, endpoint returns to an InService status. This transitional status only applies to an endpoint that has autoscaling enabled and is undergoing variant weight or capacity changes as part of an UpdateEndpointWeightsAndCapacities call or when the UpdateEndpointWeightsAndCapacities operation is called explicitly.

InService: Endpoint is available to process incoming requests.

Deleting: DeleteEndpoint is executing.

Failed: Endpoint could not be created, updated, or re-scaled. Use DescribeEndpointOutput$FailureReason for information about the failure. DeleteEndpoint is the only operation that can be performed on a failed endpoint.

describeEndpointResponse_creationTime :: Lens' DescribeEndpointResponse UTCTime

A timestamp that shows when the endpoint was created.

describeEndpointResponse_lastModifiedTime :: Lens' DescribeEndpointResponse UTCTime

A timestamp that shows when the endpoint was last modified.