Network.AWS.SageMaker.DescribeEndpointConfig

Description

Returns the description of an endpoint configuration created using the CreateEndpointConfig API.

Creating a Request

data DescribeEndpointConfig

See: newDescribeEndpointConfig smart constructor.

Constructors

DescribeEndpointConfig'

Fields

endpointConfigName :: Text

The name of the endpoint configuration.

newDescribeEndpointConfig

Arguments

:: Text
$sel:endpointConfigName:DescribeEndpointConfig'

-> DescribeEndpointConfig

Create a value of DescribeEndpointConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpointConfigName:DescribeEndpointConfig', describeEndpointConfig_endpointConfigName - The name of the endpoint configuration.

Request Lenses

describeEndpointConfig_endpointConfigName :: Lens' DescribeEndpointConfig Text

The name of the endpoint configuration.

Destructuring the Response

data DescribeEndpointConfigResponse

See: newDescribeEndpointConfigResponse smart constructor.

Constructors

DescribeEndpointConfigResponse'

Fields

asyncInferenceConfig :: Maybe AsyncInferenceConfig

Returns the description of an endpoint configuration created using the CreateEndpointConfig API.

kmsKeyId :: Maybe Text

Amazon Web Services KMS key ID Amazon SageMaker uses to encrypt data when storing it on the ML storage volume attached to the instance.

dataCaptureConfig :: Maybe DataCaptureConfig

httpStatus :: Int

The response's http status code.

endpointConfigName :: Text

Name of the Amazon SageMaker endpoint configuration.

endpointConfigArn :: Text

The Amazon Resource Name (ARN) of the endpoint configuration.

productionVariants :: NonEmpty ProductionVariant

An array of ProductionVariant objects, one for each model that you want to host at this endpoint.

creationTime :: POSIX

A timestamp that shows when the endpoint configuration was created.

newDescribeEndpointConfigResponse

Arguments

:: Int
$sel:httpStatus:DescribeEndpointConfigResponse'

-> Text
$sel:endpointConfigName:DescribeEndpointConfig'

-> Text
$sel:endpointConfigArn:DescribeEndpointConfigResponse'

-> NonEmpty ProductionVariant
$sel:productionVariants:DescribeEndpointConfigResponse'

-> UTCTime
$sel:creationTime:DescribeEndpointConfigResponse'

-> DescribeEndpointConfigResponse

Create a value of DescribeEndpointConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:asyncInferenceConfig:DescribeEndpointConfigResponse', describeEndpointConfigResponse_asyncInferenceConfig - Returns the description of an endpoint configuration created using the CreateEndpointConfig API.

$sel:kmsKeyId:DescribeEndpointConfigResponse', describeEndpointConfigResponse_kmsKeyId - Amazon Web Services KMS key ID Amazon SageMaker uses to encrypt data when storing it on the ML storage volume attached to the instance.

$sel:dataCaptureConfig:DescribeEndpointConfigResponse', describeEndpointConfigResponse_dataCaptureConfig - Undocumented member.

$sel:httpStatus:DescribeEndpointConfigResponse', describeEndpointConfigResponse_httpStatus - The response's http status code.

$sel:endpointConfigName:DescribeEndpointConfig', describeEndpointConfigResponse_endpointConfigName - Name of the Amazon SageMaker endpoint configuration.

$sel:endpointConfigArn:DescribeEndpointConfigResponse', describeEndpointConfigResponse_endpointConfigArn - The Amazon Resource Name (ARN) of the endpoint configuration.

$sel:productionVariants:DescribeEndpointConfigResponse', describeEndpointConfigResponse_productionVariants - An array of ProductionVariant objects, one for each model that you want to host at this endpoint.

$sel:creationTime:DescribeEndpointConfigResponse', describeEndpointConfigResponse_creationTime - A timestamp that shows when the endpoint configuration was created.

Response Lenses

describeEndpointConfigResponse_asyncInferenceConfig :: Lens' DescribeEndpointConfigResponse (Maybe AsyncInferenceConfig)

Returns the description of an endpoint configuration created using the CreateEndpointConfig API.

describeEndpointConfigResponse_kmsKeyId :: Lens' DescribeEndpointConfigResponse (Maybe Text)

Amazon Web Services KMS key ID Amazon SageMaker uses to encrypt data when storing it on the ML storage volume attached to the instance.

describeEndpointConfigResponse_dataCaptureConfig :: Lens' DescribeEndpointConfigResponse (Maybe DataCaptureConfig)

Undocumented member.

describeEndpointConfigResponse_httpStatus :: Lens' DescribeEndpointConfigResponse Int

The response's http status code.

describeEndpointConfigResponse_endpointConfigName :: Lens' DescribeEndpointConfigResponse Text

Name of the Amazon SageMaker endpoint configuration.

describeEndpointConfigResponse_endpointConfigArn :: Lens' DescribeEndpointConfigResponse Text

The Amazon Resource Name (ARN) of the endpoint configuration.

describeEndpointConfigResponse_productionVariants :: Lens' DescribeEndpointConfigResponse (NonEmpty ProductionVariant)

An array of ProductionVariant objects, one for each model that you want to host at this endpoint.

describeEndpointConfigResponse_creationTime :: Lens' DescribeEndpointConfigResponse UTCTime

A timestamp that shows when the endpoint configuration was created.