Network.AWS.SageMaker.DescribeExperiment

Description

Provides a list of an experiment's properties.

Creating a Request

data DescribeExperiment

See: newDescribeExperiment smart constructor.

Constructors

DescribeExperiment'

Fields

experimentName :: Text

The name of the experiment to describe.

newDescribeExperiment

Arguments

:: Text
$sel:experimentName:DescribeExperiment'

-> DescribeExperiment

Create a value of DescribeExperiment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:experimentName:DescribeExperiment', describeExperiment_experimentName - The name of the experiment to describe.

Request Lenses

describeExperiment_experimentName :: Lens' DescribeExperiment Text

The name of the experiment to describe.

Destructuring the Response

data DescribeExperimentResponse

See: newDescribeExperimentResponse smart constructor.

Constructors

DescribeExperimentResponse'

Fields

creationTime :: Maybe POSIX

When the experiment was created.

createdBy :: Maybe UserContext

Who created the experiment.

lastModifiedTime :: Maybe POSIX

When the experiment was last modified.

experimentName :: Maybe Text

The name of the experiment.

experimentArn :: Maybe Text

The Amazon Resource Name (ARN) of the experiment.

source :: Maybe ExperimentSource

The ARN of the source and, optionally, the type.

displayName :: Maybe Text

The name of the experiment as displayed. If DisplayName isn't specified, ExperimentName is displayed.

lastModifiedBy :: Maybe UserContext

Who last modified the experiment.

description :: Maybe Text

The description of the experiment.

httpStatus :: Int

The response's http status code.

newDescribeExperimentResponse

Arguments

:: Int
$sel:httpStatus:DescribeExperimentResponse'

-> DescribeExperimentResponse

Create a value of DescribeExperimentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeExperimentResponse', describeExperimentResponse_creationTime - When the experiment was created.

$sel:createdBy:DescribeExperimentResponse', describeExperimentResponse_createdBy - Who created the experiment.

$sel:lastModifiedTime:DescribeExperimentResponse', describeExperimentResponse_lastModifiedTime - When the experiment was last modified.

$sel:experimentName:DescribeExperiment', describeExperimentResponse_experimentName - The name of the experiment.

$sel:experimentArn:DescribeExperimentResponse', describeExperimentResponse_experimentArn - The Amazon Resource Name (ARN) of the experiment.

$sel:source:DescribeExperimentResponse', describeExperimentResponse_source - The ARN of the source and, optionally, the type.

$sel:displayName:DescribeExperimentResponse', describeExperimentResponse_displayName - The name of the experiment as displayed. If DisplayName isn't specified, ExperimentName is displayed.

$sel:lastModifiedBy:DescribeExperimentResponse', describeExperimentResponse_lastModifiedBy - Who last modified the experiment.

$sel:description:DescribeExperimentResponse', describeExperimentResponse_description - The description of the experiment.

$sel:httpStatus:DescribeExperimentResponse', describeExperimentResponse_httpStatus - The response's http status code.

Response Lenses

describeExperimentResponse_creationTime :: Lens' DescribeExperimentResponse (Maybe UTCTime)

When the experiment was created.

describeExperimentResponse_createdBy :: Lens' DescribeExperimentResponse (Maybe UserContext)

Who created the experiment.

describeExperimentResponse_lastModifiedTime :: Lens' DescribeExperimentResponse (Maybe UTCTime)

When the experiment was last modified.

describeExperimentResponse_experimentName :: Lens' DescribeExperimentResponse (Maybe Text)

The name of the experiment.

describeExperimentResponse_experimentArn :: Lens' DescribeExperimentResponse (Maybe Text)

The Amazon Resource Name (ARN) of the experiment.

describeExperimentResponse_source :: Lens' DescribeExperimentResponse (Maybe ExperimentSource)

The ARN of the source and, optionally, the type.

describeExperimentResponse_displayName :: Lens' DescribeExperimentResponse (Maybe Text)

The name of the experiment as displayed. If DisplayName isn't specified, ExperimentName is displayed.

describeExperimentResponse_lastModifiedBy :: Lens' DescribeExperimentResponse (Maybe UserContext)

Who last modified the experiment.

describeExperimentResponse_description :: Lens' DescribeExperimentResponse (Maybe Text)

The description of the experiment.

describeExperimentResponse_httpStatus :: Lens' DescribeExperimentResponse Int

The response's http status code.