Network.AWS.SageMaker.DescribeFeatureGroup

Description

Use this operation to describe a FeatureGroup. The response includes information on the creation time, FeatureGroup name, the unique identifier for each FeatureGroup, and more.

Creating a Request

data DescribeFeatureGroup

See: newDescribeFeatureGroup smart constructor.

Constructors

DescribeFeatureGroup'

Fields

nextToken :: Maybe Text

A token to resume pagination of the list of Features (FeatureDefinitions). 2,500 Features are returned by default.

featureGroupName :: Text

The name of the FeatureGroup you want described.

newDescribeFeatureGroup

Arguments

:: Text
$sel:featureGroupName:DescribeFeatureGroup'

-> DescribeFeatureGroup

Create a value of DescribeFeatureGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeFeatureGroup', describeFeatureGroup_nextToken - A token to resume pagination of the list of Features (FeatureDefinitions). 2,500 Features are returned by default.

$sel:featureGroupName:DescribeFeatureGroup', describeFeatureGroup_featureGroupName - The name of the FeatureGroup you want described.

Request Lenses

describeFeatureGroup_nextToken :: Lens' DescribeFeatureGroup (Maybe Text)

A token to resume pagination of the list of Features (FeatureDefinitions). 2,500 Features are returned by default.

describeFeatureGroup_featureGroupName :: Lens' DescribeFeatureGroup Text

The name of the FeatureGroup you want described.

Destructuring the Response

data DescribeFeatureGroupResponse

See: newDescribeFeatureGroupResponse smart constructor.

Constructors

DescribeFeatureGroupResponse'

Fields

offlineStoreConfig :: Maybe OfflineStoreConfig

The configuration of the OfflineStore, inducing the S3 location of the OfflineStore, Amazon Web Services Glue or Amazon Web Services Hive data catalogue configurations, and the security configuration.

failureReason :: Maybe Text

The reason that the FeatureGroup failed to be replicated in the OfflineStore. This is failure can occur because:

The FeatureGroup could not be created in the OfflineStore.

The FeatureGroup could not be deleted from the OfflineStore.

featureGroupStatus :: Maybe FeatureGroupStatus

The status of the feature group.

offlineStoreStatus :: Maybe OfflineStoreStatus

The status of the OfflineStore. Notifies you if replicating data into the OfflineStore has failed. Returns either: Active or Blocked

onlineStoreConfig :: Maybe OnlineStoreConfig

The configuration for the OnlineStore.

description :: Maybe Text

A free form description of the feature group.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM execution role used to persist data into the OfflineStore if an OfflineStoreConfig is provided.

httpStatus :: Int

The response's http status code.

featureGroupArn :: Text

The Amazon Resource Name (ARN) of the FeatureGroup.

featureGroupName :: Text

he name of the FeatureGroup.

recordIdentifierFeatureName :: Text

The name of the Feature used for RecordIdentifier, whose value uniquely identifies a record stored in the feature store.

eventTimeFeatureName :: Text

The name of the feature that stores the EventTime of a Record in a FeatureGroup.

An EventTime is a point in time when a new event occurs that corresponds to the creation or update of a Record in a FeatureGroup. All Records in the FeatureGroup have a corresponding EventTime.

featureDefinitions :: NonEmpty FeatureDefinition

A list of the Features in the FeatureGroup. Each feature is defined by a FeatureName and FeatureType.

creationTime :: POSIX

A timestamp indicating when SageMaker created the FeatureGroup.

nextToken :: Text

A token to resume pagination of the list of Features (FeatureDefinitions).

newDescribeFeatureGroupResponse

Arguments

:: Int
$sel:httpStatus:DescribeFeatureGroupResponse'

-> Text
$sel:featureGroupArn:DescribeFeatureGroupResponse'

-> Text
$sel:featureGroupName:DescribeFeatureGroup'

-> Text
$sel:recordIdentifierFeatureName:DescribeFeatureGroupResponse'

-> Text
$sel:eventTimeFeatureName:DescribeFeatureGroupResponse'

-> NonEmpty FeatureDefinition
$sel:featureDefinitions:DescribeFeatureGroupResponse'

-> UTCTime
$sel:creationTime:DescribeFeatureGroupResponse'

-> Text
$sel:nextToken:DescribeFeatureGroup'

-> DescribeFeatureGroupResponse

Create a value of DescribeFeatureGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:offlineStoreConfig:DescribeFeatureGroupResponse', describeFeatureGroupResponse_offlineStoreConfig - The configuration of the OfflineStore, inducing the S3 location of the OfflineStore, Amazon Web Services Glue or Amazon Web Services Hive data catalogue configurations, and the security configuration.

$sel:failureReason:DescribeFeatureGroupResponse', describeFeatureGroupResponse_failureReason - The reason that the FeatureGroup failed to be replicated in the OfflineStore. This is failure can occur because:

The FeatureGroup could not be created in the OfflineStore.

The FeatureGroup could not be deleted from the OfflineStore.

$sel:featureGroupStatus:DescribeFeatureGroupResponse', describeFeatureGroupResponse_featureGroupStatus - The status of the feature group.

$sel:offlineStoreStatus:DescribeFeatureGroupResponse', describeFeatureGroupResponse_offlineStoreStatus - The status of the OfflineStore. Notifies you if replicating data into the OfflineStore has failed. Returns either: Active or Blocked

$sel:onlineStoreConfig:DescribeFeatureGroupResponse', describeFeatureGroupResponse_onlineStoreConfig - The configuration for the OnlineStore.

$sel:description:DescribeFeatureGroupResponse', describeFeatureGroupResponse_description - A free form description of the feature group.

$sel:roleArn:DescribeFeatureGroupResponse', describeFeatureGroupResponse_roleArn - The Amazon Resource Name (ARN) of the IAM execution role used to persist data into the OfflineStore if an OfflineStoreConfig is provided.

$sel:httpStatus:DescribeFeatureGroupResponse', describeFeatureGroupResponse_httpStatus - The response's http status code.

$sel:featureGroupArn:DescribeFeatureGroupResponse', describeFeatureGroupResponse_featureGroupArn - The Amazon Resource Name (ARN) of the FeatureGroup.

$sel:featureGroupName:DescribeFeatureGroup', describeFeatureGroupResponse_featureGroupName - he name of the FeatureGroup.

$sel:recordIdentifierFeatureName:DescribeFeatureGroupResponse', describeFeatureGroupResponse_recordIdentifierFeatureName - The name of the Feature used for RecordIdentifier, whose value uniquely identifies a record stored in the feature store.

$sel:eventTimeFeatureName:DescribeFeatureGroupResponse', describeFeatureGroupResponse_eventTimeFeatureName - The name of the feature that stores the EventTime of a Record in a FeatureGroup.

An EventTime is a point in time when a new event occurs that corresponds to the creation or update of a Record in a FeatureGroup. All Records in the FeatureGroup have a corresponding EventTime.

$sel:featureDefinitions:DescribeFeatureGroupResponse', describeFeatureGroupResponse_featureDefinitions - A list of the Features in the FeatureGroup. Each feature is defined by a FeatureName and FeatureType.

$sel:creationTime:DescribeFeatureGroupResponse', describeFeatureGroupResponse_creationTime - A timestamp indicating when SageMaker created the FeatureGroup.

$sel:nextToken:DescribeFeatureGroup', describeFeatureGroupResponse_nextToken - A token to resume pagination of the list of Features (FeatureDefinitions).

Response Lenses

describeFeatureGroupResponse_offlineStoreConfig :: Lens' DescribeFeatureGroupResponse (Maybe OfflineStoreConfig)

The configuration of the OfflineStore, inducing the S3 location of the OfflineStore, Amazon Web Services Glue or Amazon Web Services Hive data catalogue configurations, and the security configuration.

describeFeatureGroupResponse_failureReason :: Lens' DescribeFeatureGroupResponse (Maybe Text)

The reason that the FeatureGroup failed to be replicated in the OfflineStore. This is failure can occur because:

The FeatureGroup could not be created in the OfflineStore.

The FeatureGroup could not be deleted from the OfflineStore.

describeFeatureGroupResponse_featureGroupStatus :: Lens' DescribeFeatureGroupResponse (Maybe FeatureGroupStatus)

The status of the feature group.

describeFeatureGroupResponse_offlineStoreStatus :: Lens' DescribeFeatureGroupResponse (Maybe OfflineStoreStatus)

The status of the OfflineStore. Notifies you if replicating data into the OfflineStore has failed. Returns either: Active or Blocked

describeFeatureGroupResponse_onlineStoreConfig :: Lens' DescribeFeatureGroupResponse (Maybe OnlineStoreConfig)

The configuration for the OnlineStore.

describeFeatureGroupResponse_description :: Lens' DescribeFeatureGroupResponse (Maybe Text)

A free form description of the feature group.

describeFeatureGroupResponse_roleArn :: Lens' DescribeFeatureGroupResponse (Maybe Text)

The Amazon Resource Name (ARN) of the IAM execution role used to persist data into the OfflineStore if an OfflineStoreConfig is provided.

describeFeatureGroupResponse_httpStatus :: Lens' DescribeFeatureGroupResponse Int

The response's http status code.

describeFeatureGroupResponse_featureGroupArn :: Lens' DescribeFeatureGroupResponse Text

The Amazon Resource Name (ARN) of the FeatureGroup.

describeFeatureGroupResponse_featureGroupName :: Lens' DescribeFeatureGroupResponse Text

he name of the FeatureGroup.

describeFeatureGroupResponse_recordIdentifierFeatureName :: Lens' DescribeFeatureGroupResponse Text

The name of the Feature used for RecordIdentifier, whose value uniquely identifies a record stored in the feature store.

describeFeatureGroupResponse_eventTimeFeatureName :: Lens' DescribeFeatureGroupResponse Text

The name of the feature that stores the EventTime of a Record in a FeatureGroup.

An EventTime is a point in time when a new event occurs that corresponds to the creation or update of a Record in a FeatureGroup. All Records in the FeatureGroup have a corresponding EventTime.

describeFeatureGroupResponse_featureDefinitions :: Lens' DescribeFeatureGroupResponse (NonEmpty FeatureDefinition)

A list of the Features in the FeatureGroup. Each feature is defined by a FeatureName and FeatureType.

describeFeatureGroupResponse_creationTime :: Lens' DescribeFeatureGroupResponse UTCTime

A timestamp indicating when SageMaker created the FeatureGroup.

describeFeatureGroupResponse_nextToken :: Lens' DescribeFeatureGroupResponse Text

A token to resume pagination of the list of Features (FeatureDefinitions).