Network.AWS.SageMaker.DescribeHyperParameterTuningJob

Description

Gets a description of a hyperparameter tuning job.

Creating a Request

data DescribeHyperParameterTuningJob

See: newDescribeHyperParameterTuningJob smart constructor.

Constructors

DescribeHyperParameterTuningJob'

Fields

hyperParameterTuningJobName :: Text

The name of the tuning job.

newDescribeHyperParameterTuningJob

Arguments

:: Text
$sel:hyperParameterTuningJobName:DescribeHyperParameterTuningJob'

-> DescribeHyperParameterTuningJob

Create a value of DescribeHyperParameterTuningJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hyperParameterTuningJobName:DescribeHyperParameterTuningJob', describeHyperParameterTuningJob_hyperParameterTuningJobName - The name of the tuning job.

Request Lenses

describeHyperParameterTuningJob_hyperParameterTuningJobName :: Lens' DescribeHyperParameterTuningJob Text

The name of the tuning job.

Destructuring the Response

data DescribeHyperParameterTuningJobResponse

See: newDescribeHyperParameterTuningJobResponse smart constructor.

Constructors

DescribeHyperParameterTuningJobResponse'

Fields

bestTrainingJob :: Maybe HyperParameterTrainingJobSummary

A TrainingJobSummary object that describes the training job that completed with the best current HyperParameterTuningJobObjective.

warmStartConfig :: Maybe HyperParameterTuningJobWarmStartConfig

The configuration for starting the hyperparameter parameter tuning job using one or more previous tuning jobs as a starting point. The results of previous tuning jobs are used to inform which combinations of hyperparameters to search over in the new tuning job.

hyperParameterTuningEndTime :: Maybe POSIX

The date and time that the tuning job ended.

failureReason :: Maybe Text

If the tuning job failed, the reason it failed.

trainingJobDefinitions :: Maybe (NonEmpty HyperParameterTrainingJobDefinition)

A list of the HyperParameterTrainingJobDefinition objects launched for this tuning job.

lastModifiedTime :: Maybe POSIX

The date and time that the status of the tuning job was modified.

overallBestTrainingJob :: Maybe HyperParameterTrainingJobSummary

If the hyperparameter tuning job is an warm start tuning job with a WarmStartType of IDENTICAL_DATA_AND_ALGORITHM, this is the TrainingJobSummary for the training job with the best objective metric value of all training jobs launched by this tuning job and all parent jobs specified for the warm start tuning job.

trainingJobDefinition :: Maybe HyperParameterTrainingJobDefinition

The HyperParameterTrainingJobDefinition object that specifies the definition of the training jobs that this tuning job launches.

httpStatus :: Int

The response's http status code.

hyperParameterTuningJobName :: Text

The name of the tuning job.

hyperParameterTuningJobArn :: Text

The Amazon Resource Name (ARN) of the tuning job.

hyperParameterTuningJobConfig :: HyperParameterTuningJobConfig

The HyperParameterTuningJobConfig object that specifies the configuration of the tuning job.

hyperParameterTuningJobStatus :: HyperParameterTuningJobStatus

The status of the tuning job: InProgress, Completed, Failed, Stopping, or Stopped.

creationTime :: POSIX

The date and time that the tuning job started.

trainingJobStatusCounters :: TrainingJobStatusCounters

The TrainingJobStatusCounters object that specifies the number of training jobs, categorized by status, that this tuning job launched.

objectiveStatusCounters :: ObjectiveStatusCounters

The ObjectiveStatusCounters object that specifies the number of training jobs, categorized by the status of their final objective metric, that this tuning job launched.

newDescribeHyperParameterTuningJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeHyperParameterTuningJobResponse'

-> Text
$sel:hyperParameterTuningJobName:DescribeHyperParameterTuningJob'

-> Text
$sel:hyperParameterTuningJobArn:DescribeHyperParameterTuningJobResponse'

-> HyperParameterTuningJobConfig
$sel:hyperParameterTuningJobConfig:DescribeHyperParameterTuningJobResponse'

-> HyperParameterTuningJobStatus
$sel:hyperParameterTuningJobStatus:DescribeHyperParameterTuningJobResponse'

-> UTCTime
$sel:creationTime:DescribeHyperParameterTuningJobResponse'

-> TrainingJobStatusCounters
$sel:trainingJobStatusCounters:DescribeHyperParameterTuningJobResponse'

-> ObjectiveStatusCounters
$sel:objectiveStatusCounters:DescribeHyperParameterTuningJobResponse'

-> DescribeHyperParameterTuningJobResponse

Create a value of DescribeHyperParameterTuningJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bestTrainingJob:DescribeHyperParameterTuningJobResponse', describeHyperParameterTuningJobResponse_bestTrainingJob - A TrainingJobSummary object that describes the training job that completed with the best current HyperParameterTuningJobObjective.

$sel:warmStartConfig:DescribeHyperParameterTuningJobResponse', describeHyperParameterTuningJobResponse_warmStartConfig - The configuration for starting the hyperparameter parameter tuning job using one or more previous tuning jobs as a starting point. The results of previous tuning jobs are used to inform which combinations of hyperparameters to search over in the new tuning job.

$sel:hyperParameterTuningEndTime:DescribeHyperParameterTuningJobResponse', describeHyperParameterTuningJobResponse_hyperParameterTuningEndTime - The date and time that the tuning job ended.

$sel:failureReason:DescribeHyperParameterTuningJobResponse', describeHyperParameterTuningJobResponse_failureReason - If the tuning job failed, the reason it failed.

$sel:trainingJobDefinitions:DescribeHyperParameterTuningJobResponse', describeHyperParameterTuningJobResponse_trainingJobDefinitions - A list of the HyperParameterTrainingJobDefinition objects launched for this tuning job.

$sel:lastModifiedTime:DescribeHyperParameterTuningJobResponse', describeHyperParameterTuningJobResponse_lastModifiedTime - The date and time that the status of the tuning job was modified.

$sel:overallBestTrainingJob:DescribeHyperParameterTuningJobResponse', describeHyperParameterTuningJobResponse_overallBestTrainingJob - If the hyperparameter tuning job is an warm start tuning job with a WarmStartType of IDENTICAL_DATA_AND_ALGORITHM, this is the TrainingJobSummary for the training job with the best objective metric value of all training jobs launched by this tuning job and all parent jobs specified for the warm start tuning job.

$sel:trainingJobDefinition:DescribeHyperParameterTuningJobResponse', describeHyperParameterTuningJobResponse_trainingJobDefinition - The HyperParameterTrainingJobDefinition object that specifies the definition of the training jobs that this tuning job launches.

$sel:httpStatus:DescribeHyperParameterTuningJobResponse', describeHyperParameterTuningJobResponse_httpStatus - The response's http status code.

$sel:hyperParameterTuningJobName:DescribeHyperParameterTuningJob', describeHyperParameterTuningJobResponse_hyperParameterTuningJobName - The name of the tuning job.

$sel:hyperParameterTuningJobArn:DescribeHyperParameterTuningJobResponse', describeHyperParameterTuningJobResponse_hyperParameterTuningJobArn - The Amazon Resource Name (ARN) of the tuning job.

$sel:hyperParameterTuningJobConfig:DescribeHyperParameterTuningJobResponse', describeHyperParameterTuningJobResponse_hyperParameterTuningJobConfig - The HyperParameterTuningJobConfig object that specifies the configuration of the tuning job.

$sel:hyperParameterTuningJobStatus:DescribeHyperParameterTuningJobResponse', describeHyperParameterTuningJobResponse_hyperParameterTuningJobStatus - The status of the tuning job: InProgress, Completed, Failed, Stopping, or Stopped.

$sel:creationTime:DescribeHyperParameterTuningJobResponse', describeHyperParameterTuningJobResponse_creationTime - The date and time that the tuning job started.

$sel:trainingJobStatusCounters:DescribeHyperParameterTuningJobResponse', describeHyperParameterTuningJobResponse_trainingJobStatusCounters - The TrainingJobStatusCounters object that specifies the number of training jobs, categorized by status, that this tuning job launched.

$sel:objectiveStatusCounters:DescribeHyperParameterTuningJobResponse', describeHyperParameterTuningJobResponse_objectiveStatusCounters - The ObjectiveStatusCounters object that specifies the number of training jobs, categorized by the status of their final objective metric, that this tuning job launched.

Response Lenses

describeHyperParameterTuningJobResponse_bestTrainingJob :: Lens' DescribeHyperParameterTuningJobResponse (Maybe HyperParameterTrainingJobSummary)

A TrainingJobSummary object that describes the training job that completed with the best current HyperParameterTuningJobObjective.

describeHyperParameterTuningJobResponse_warmStartConfig :: Lens' DescribeHyperParameterTuningJobResponse (Maybe HyperParameterTuningJobWarmStartConfig)

The configuration for starting the hyperparameter parameter tuning job using one or more previous tuning jobs as a starting point. The results of previous tuning jobs are used to inform which combinations of hyperparameters to search over in the new tuning job.

describeHyperParameterTuningJobResponse_hyperParameterTuningEndTime :: Lens' DescribeHyperParameterTuningJobResponse (Maybe UTCTime)

The date and time that the tuning job ended.

describeHyperParameterTuningJobResponse_failureReason :: Lens' DescribeHyperParameterTuningJobResponse (Maybe Text)

If the tuning job failed, the reason it failed.

describeHyperParameterTuningJobResponse_trainingJobDefinitions :: Lens' DescribeHyperParameterTuningJobResponse (Maybe (NonEmpty HyperParameterTrainingJobDefinition))

A list of the HyperParameterTrainingJobDefinition objects launched for this tuning job.

describeHyperParameterTuningJobResponse_lastModifiedTime :: Lens' DescribeHyperParameterTuningJobResponse (Maybe UTCTime)

The date and time that the status of the tuning job was modified.

describeHyperParameterTuningJobResponse_overallBestTrainingJob :: Lens' DescribeHyperParameterTuningJobResponse (Maybe HyperParameterTrainingJobSummary)

If the hyperparameter tuning job is an warm start tuning job with a WarmStartType of IDENTICAL_DATA_AND_ALGORITHM, this is the TrainingJobSummary for the training job with the best objective metric value of all training jobs launched by this tuning job and all parent jobs specified for the warm start tuning job.

describeHyperParameterTuningJobResponse_trainingJobDefinition :: Lens' DescribeHyperParameterTuningJobResponse (Maybe HyperParameterTrainingJobDefinition)

The HyperParameterTrainingJobDefinition object that specifies the definition of the training jobs that this tuning job launches.

describeHyperParameterTuningJobResponse_httpStatus :: Lens' DescribeHyperParameterTuningJobResponse Int

The response's http status code.

describeHyperParameterTuningJobResponse_hyperParameterTuningJobName :: Lens' DescribeHyperParameterTuningJobResponse Text

The name of the tuning job.

describeHyperParameterTuningJobResponse_hyperParameterTuningJobArn :: Lens' DescribeHyperParameterTuningJobResponse Text

The Amazon Resource Name (ARN) of the tuning job.

describeHyperParameterTuningJobResponse_hyperParameterTuningJobConfig :: Lens' DescribeHyperParameterTuningJobResponse HyperParameterTuningJobConfig

The HyperParameterTuningJobConfig object that specifies the configuration of the tuning job.

describeHyperParameterTuningJobResponse_hyperParameterTuningJobStatus :: Lens' DescribeHyperParameterTuningJobResponse HyperParameterTuningJobStatus

The status of the tuning job: InProgress, Completed, Failed, Stopping, or Stopped.

describeHyperParameterTuningJobResponse_creationTime :: Lens' DescribeHyperParameterTuningJobResponse UTCTime

The date and time that the tuning job started.

describeHyperParameterTuningJobResponse_trainingJobStatusCounters :: Lens' DescribeHyperParameterTuningJobResponse TrainingJobStatusCounters

The TrainingJobStatusCounters object that specifies the number of training jobs, categorized by status, that this tuning job launched.

describeHyperParameterTuningJobResponse_objectiveStatusCounters :: Lens' DescribeHyperParameterTuningJobResponse ObjectiveStatusCounters

The ObjectiveStatusCounters object that specifies the number of training jobs, categorized by the status of their final objective metric, that this tuning job launched.