Network.AWS.SageMaker.DescribeImage

Description

Describes a SageMaker image.

Creating a Request

data DescribeImage

See: newDescribeImage smart constructor.

Constructors

DescribeImage'

Fields

imageName :: Text

The name of the image to describe.

newDescribeImage

Arguments

:: Text
$sel:imageName:DescribeImage'

-> DescribeImage

Create a value of DescribeImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageName:DescribeImage', describeImage_imageName - The name of the image to describe.

Request Lenses

describeImage_imageName :: Lens' DescribeImage Text

The name of the image to describe.

Destructuring the Response

data DescribeImageResponse

See: newDescribeImageResponse smart constructor.

Constructors

DescribeImageResponse'

Fields

creationTime :: Maybe POSIX

When the image was created.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role that enables Amazon SageMaker to perform tasks on your behalf.

imageName :: Maybe Text

The name of the image.

failureReason :: Maybe Text

When a create, update, or delete operation fails, the reason for the failure.

lastModifiedTime :: Maybe POSIX

When the image was last modified.

description :: Maybe Text

The description of the image.

imageStatus :: Maybe ImageStatus

The status of the image.

displayName :: Maybe Text

The name of the image as displayed.

imageArn :: Maybe Text

The Amazon Resource Name (ARN) of the image.

httpStatus :: Int

The response's http status code.

newDescribeImageResponse

Arguments

:: Int
$sel:httpStatus:DescribeImageResponse'

-> DescribeImageResponse

Create a value of DescribeImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeImageResponse', describeImageResponse_creationTime - When the image was created.

$sel:roleArn:DescribeImageResponse', describeImageResponse_roleArn - The Amazon Resource Name (ARN) of the IAM role that enables Amazon SageMaker to perform tasks on your behalf.

$sel:imageName:DescribeImage', describeImageResponse_imageName - The name of the image.

$sel:failureReason:DescribeImageResponse', describeImageResponse_failureReason - When a create, update, or delete operation fails, the reason for the failure.

$sel:lastModifiedTime:DescribeImageResponse', describeImageResponse_lastModifiedTime - When the image was last modified.

$sel:description:DescribeImageResponse', describeImageResponse_description - The description of the image.

$sel:imageStatus:DescribeImageResponse', describeImageResponse_imageStatus - The status of the image.

$sel:displayName:DescribeImageResponse', describeImageResponse_displayName - The name of the image as displayed.

$sel:imageArn:DescribeImageResponse', describeImageResponse_imageArn - The Amazon Resource Name (ARN) of the image.

$sel:httpStatus:DescribeImageResponse', describeImageResponse_httpStatus - The response's http status code.

Response Lenses

describeImageResponse_creationTime :: Lens' DescribeImageResponse (Maybe UTCTime)

When the image was created.

describeImageResponse_roleArn :: Lens' DescribeImageResponse (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that enables Amazon SageMaker to perform tasks on your behalf.

describeImageResponse_imageName :: Lens' DescribeImageResponse (Maybe Text)

The name of the image.

describeImageResponse_failureReason :: Lens' DescribeImageResponse (Maybe Text)

When a create, update, or delete operation fails, the reason for the failure.

describeImageResponse_lastModifiedTime :: Lens' DescribeImageResponse (Maybe UTCTime)

When the image was last modified.

describeImageResponse_description :: Lens' DescribeImageResponse (Maybe Text)

The description of the image.

describeImageResponse_imageStatus :: Lens' DescribeImageResponse (Maybe ImageStatus)

The status of the image.

describeImageResponse_displayName :: Lens' DescribeImageResponse (Maybe Text)

The name of the image as displayed.

describeImageResponse_imageArn :: Lens' DescribeImageResponse (Maybe Text)

The Amazon Resource Name (ARN) of the image.

describeImageResponse_httpStatus :: Lens' DescribeImageResponse Int

The response's http status code.