Network.AWS.SageMaker.DescribeModelBiasJobDefinition

Description

Returns a description of a model bias job definition.

Creating a Request

data DescribeModelBiasJobDefinition

See: newDescribeModelBiasJobDefinition smart constructor.

Constructors

DescribeModelBiasJobDefinition'

Fields

jobDefinitionName :: Text

The name of the model bias job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

newDescribeModelBiasJobDefinition

Arguments

:: Text
$sel:jobDefinitionName:DescribeModelBiasJobDefinition'

-> DescribeModelBiasJobDefinition

Create a value of DescribeModelBiasJobDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobDefinitionName:DescribeModelBiasJobDefinition', describeModelBiasJobDefinition_jobDefinitionName - The name of the model bias job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

Request Lenses

describeModelBiasJobDefinition_jobDefinitionName :: Lens' DescribeModelBiasJobDefinition Text

The name of the model bias job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

Destructuring the Response

data DescribeModelBiasJobDefinitionResponse

See: newDescribeModelBiasJobDefinitionResponse smart constructor.

Constructors

DescribeModelBiasJobDefinitionResponse'

Fields

networkConfig :: Maybe MonitoringNetworkConfig

Networking options for a model bias job.

modelBiasBaselineConfig :: Maybe ModelBiasBaselineConfig

The baseline configuration for a model bias job.

stoppingCondition :: Maybe MonitoringStoppingCondition

httpStatus :: Int

The response's http status code.

jobDefinitionArn :: Text

The Amazon Resource Name (ARN) of the model bias job.

jobDefinitionName :: Text

The name of the bias job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

creationTime :: POSIX

The time at which the model bias job was created.

modelBiasAppSpecification :: ModelBiasAppSpecification

Configures the model bias job to run a specified Docker container image.

modelBiasJobInput :: ModelBiasJobInput

Inputs for the model bias job.

modelBiasJobOutputConfig :: MonitoringOutputConfig

jobResources :: MonitoringResources

roleArn :: Text

The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that has read permission to the input data location and write permission to the output data location in Amazon S3.

newDescribeModelBiasJobDefinitionResponse

Arguments

:: Int
$sel:httpStatus:DescribeModelBiasJobDefinitionResponse'

-> Text
$sel:jobDefinitionArn:DescribeModelBiasJobDefinitionResponse'

-> Text
$sel:jobDefinitionName:DescribeModelBiasJobDefinition'

-> UTCTime
$sel:creationTime:DescribeModelBiasJobDefinitionResponse'

-> ModelBiasAppSpecification
$sel:modelBiasAppSpecification:DescribeModelBiasJobDefinitionResponse'

-> ModelBiasJobInput
$sel:modelBiasJobInput:DescribeModelBiasJobDefinitionResponse'

-> MonitoringOutputConfig
$sel:modelBiasJobOutputConfig:DescribeModelBiasJobDefinitionResponse'

-> MonitoringResources
$sel:jobResources:DescribeModelBiasJobDefinitionResponse'

-> Text
$sel:roleArn:DescribeModelBiasJobDefinitionResponse'

-> DescribeModelBiasJobDefinitionResponse

Create a value of DescribeModelBiasJobDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkConfig:DescribeModelBiasJobDefinitionResponse', describeModelBiasJobDefinitionResponse_networkConfig - Networking options for a model bias job.

$sel:modelBiasBaselineConfig:DescribeModelBiasJobDefinitionResponse', describeModelBiasJobDefinitionResponse_modelBiasBaselineConfig - The baseline configuration for a model bias job.

$sel:stoppingCondition:DescribeModelBiasJobDefinitionResponse', describeModelBiasJobDefinitionResponse_stoppingCondition - Undocumented member.

$sel:httpStatus:DescribeModelBiasJobDefinitionResponse', describeModelBiasJobDefinitionResponse_httpStatus - The response's http status code.

$sel:jobDefinitionArn:DescribeModelBiasJobDefinitionResponse', describeModelBiasJobDefinitionResponse_jobDefinitionArn - The Amazon Resource Name (ARN) of the model bias job.

$sel:jobDefinitionName:DescribeModelBiasJobDefinition', describeModelBiasJobDefinitionResponse_jobDefinitionName - The name of the bias job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

$sel:creationTime:DescribeModelBiasJobDefinitionResponse', describeModelBiasJobDefinitionResponse_creationTime - The time at which the model bias job was created.

$sel:modelBiasAppSpecification:DescribeModelBiasJobDefinitionResponse', describeModelBiasJobDefinitionResponse_modelBiasAppSpecification - Configures the model bias job to run a specified Docker container image.

$sel:modelBiasJobInput:DescribeModelBiasJobDefinitionResponse', describeModelBiasJobDefinitionResponse_modelBiasJobInput - Inputs for the model bias job.

$sel:modelBiasJobOutputConfig:DescribeModelBiasJobDefinitionResponse', describeModelBiasJobDefinitionResponse_modelBiasJobOutputConfig - Undocumented member.

$sel:jobResources:DescribeModelBiasJobDefinitionResponse', describeModelBiasJobDefinitionResponse_jobResources - Undocumented member.

$sel:roleArn:DescribeModelBiasJobDefinitionResponse', describeModelBiasJobDefinitionResponse_roleArn - The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that has read permission to the input data location and write permission to the output data location in Amazon S3.

Response Lenses

describeModelBiasJobDefinitionResponse_networkConfig :: Lens' DescribeModelBiasJobDefinitionResponse (Maybe MonitoringNetworkConfig)

Networking options for a model bias job.

describeModelBiasJobDefinitionResponse_modelBiasBaselineConfig :: Lens' DescribeModelBiasJobDefinitionResponse (Maybe ModelBiasBaselineConfig)

The baseline configuration for a model bias job.

describeModelBiasJobDefinitionResponse_stoppingCondition :: Lens' DescribeModelBiasJobDefinitionResponse (Maybe MonitoringStoppingCondition)

Undocumented member.

describeModelBiasJobDefinitionResponse_httpStatus :: Lens' DescribeModelBiasJobDefinitionResponse Int

The response's http status code.

describeModelBiasJobDefinitionResponse_jobDefinitionArn :: Lens' DescribeModelBiasJobDefinitionResponse Text

The Amazon Resource Name (ARN) of the model bias job.

describeModelBiasJobDefinitionResponse_jobDefinitionName :: Lens' DescribeModelBiasJobDefinitionResponse Text

The name of the bias job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

describeModelBiasJobDefinitionResponse_creationTime :: Lens' DescribeModelBiasJobDefinitionResponse UTCTime

The time at which the model bias job was created.

describeModelBiasJobDefinitionResponse_modelBiasAppSpecification :: Lens' DescribeModelBiasJobDefinitionResponse ModelBiasAppSpecification

Configures the model bias job to run a specified Docker container image.

describeModelBiasJobDefinitionResponse_modelBiasJobInput :: Lens' DescribeModelBiasJobDefinitionResponse ModelBiasJobInput

Inputs for the model bias job.

describeModelBiasJobDefinitionResponse_modelBiasJobOutputConfig :: Lens' DescribeModelBiasJobDefinitionResponse MonitoringOutputConfig

Undocumented member.

describeModelBiasJobDefinitionResponse_jobResources :: Lens' DescribeModelBiasJobDefinitionResponse MonitoringResources

Undocumented member.

describeModelBiasJobDefinitionResponse_roleArn :: Lens' DescribeModelBiasJobDefinitionResponse Text

The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that has read permission to the input data location and write permission to the output data location in Amazon S3.