Network.AWS.SageMaker.DescribeModelExplainabilityJobDefinition

Description

Returns a description of a model explainability job definition.

Creating a Request

data DescribeModelExplainabilityJobDefinition

See: newDescribeModelExplainabilityJobDefinition smart constructor.

Constructors

DescribeModelExplainabilityJobDefinition'

Fields

jobDefinitionName :: Text

The name of the model explainability job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

newDescribeModelExplainabilityJobDefinition

Arguments

:: Text
$sel:jobDefinitionName:DescribeModelExplainabilityJobDefinition'

-> DescribeModelExplainabilityJobDefinition

Create a value of DescribeModelExplainabilityJobDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobDefinitionName:DescribeModelExplainabilityJobDefinition', describeModelExplainabilityJobDefinition_jobDefinitionName - The name of the model explainability job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

Request Lenses

describeModelExplainabilityJobDefinition_jobDefinitionName :: Lens' DescribeModelExplainabilityJobDefinition Text

The name of the model explainability job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

Destructuring the Response

data DescribeModelExplainabilityJobDefinitionResponse

See: newDescribeModelExplainabilityJobDefinitionResponse smart constructor.

Constructors

DescribeModelExplainabilityJobDefinitionResponse'

Fields

networkConfig :: Maybe MonitoringNetworkConfig

Networking options for a model explainability job.

modelExplainabilityBaselineConfig :: Maybe ModelExplainabilityBaselineConfig

The baseline configuration for a model explainability job.

stoppingCondition :: Maybe MonitoringStoppingCondition

httpStatus :: Int

The response's http status code.

jobDefinitionArn :: Text

The Amazon Resource Name (ARN) of the model explainability job.

jobDefinitionName :: Text

The name of the explainability job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

creationTime :: POSIX

The time at which the model explainability job was created.

modelExplainabilityAppSpecification :: ModelExplainabilityAppSpecification

Configures the model explainability job to run a specified Docker container image.

modelExplainabilityJobInput :: ModelExplainabilityJobInput

Inputs for the model explainability job.

modelExplainabilityJobOutputConfig :: MonitoringOutputConfig

jobResources :: MonitoringResources

roleArn :: Text

The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that has read permission to the input data location and write permission to the output data location in Amazon S3.

newDescribeModelExplainabilityJobDefinitionResponse

Arguments

:: Int
$sel:httpStatus:DescribeModelExplainabilityJobDefinitionResponse'

-> Text
$sel:jobDefinitionArn:DescribeModelExplainabilityJobDefinitionResponse'

-> Text
$sel:jobDefinitionName:DescribeModelExplainabilityJobDefinition'

-> UTCTime
$sel:creationTime:DescribeModelExplainabilityJobDefinitionResponse'

-> ModelExplainabilityAppSpecification
$sel:modelExplainabilityAppSpecification:DescribeModelExplainabilityJobDefinitionResponse'

-> ModelExplainabilityJobInput
$sel:modelExplainabilityJobInput:DescribeModelExplainabilityJobDefinitionResponse'

-> MonitoringOutputConfig
$sel:modelExplainabilityJobOutputConfig:DescribeModelExplainabilityJobDefinitionResponse'

-> MonitoringResources
$sel:jobResources:DescribeModelExplainabilityJobDefinitionResponse'

-> Text
$sel:roleArn:DescribeModelExplainabilityJobDefinitionResponse'

-> DescribeModelExplainabilityJobDefinitionResponse

Create a value of DescribeModelExplainabilityJobDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkConfig:DescribeModelExplainabilityJobDefinitionResponse', describeModelExplainabilityJobDefinitionResponse_networkConfig - Networking options for a model explainability job.

$sel:modelExplainabilityBaselineConfig:DescribeModelExplainabilityJobDefinitionResponse', describeModelExplainabilityJobDefinitionResponse_modelExplainabilityBaselineConfig - The baseline configuration for a model explainability job.

$sel:stoppingCondition:DescribeModelExplainabilityJobDefinitionResponse', describeModelExplainabilityJobDefinitionResponse_stoppingCondition - Undocumented member.

$sel:httpStatus:DescribeModelExplainabilityJobDefinitionResponse', describeModelExplainabilityJobDefinitionResponse_httpStatus - The response's http status code.

$sel:jobDefinitionArn:DescribeModelExplainabilityJobDefinitionResponse', describeModelExplainabilityJobDefinitionResponse_jobDefinitionArn - The Amazon Resource Name (ARN) of the model explainability job.

$sel:jobDefinitionName:DescribeModelExplainabilityJobDefinition', describeModelExplainabilityJobDefinitionResponse_jobDefinitionName - The name of the explainability job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

$sel:creationTime:DescribeModelExplainabilityJobDefinitionResponse', describeModelExplainabilityJobDefinitionResponse_creationTime - The time at which the model explainability job was created.

$sel:modelExplainabilityAppSpecification:DescribeModelExplainabilityJobDefinitionResponse', describeModelExplainabilityJobDefinitionResponse_modelExplainabilityAppSpecification - Configures the model explainability job to run a specified Docker container image.

$sel:modelExplainabilityJobInput:DescribeModelExplainabilityJobDefinitionResponse', describeModelExplainabilityJobDefinitionResponse_modelExplainabilityJobInput - Inputs for the model explainability job.

$sel:modelExplainabilityJobOutputConfig:DescribeModelExplainabilityJobDefinitionResponse', describeModelExplainabilityJobDefinitionResponse_modelExplainabilityJobOutputConfig - Undocumented member.

$sel:jobResources:DescribeModelExplainabilityJobDefinitionResponse', describeModelExplainabilityJobDefinitionResponse_jobResources - Undocumented member.

$sel:roleArn:DescribeModelExplainabilityJobDefinitionResponse', describeModelExplainabilityJobDefinitionResponse_roleArn - The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that has read permission to the input data location and write permission to the output data location in Amazon S3.

Response Lenses

describeModelExplainabilityJobDefinitionResponse_networkConfig :: Lens' DescribeModelExplainabilityJobDefinitionResponse (Maybe MonitoringNetworkConfig)

Networking options for a model explainability job.

describeModelExplainabilityJobDefinitionResponse_modelExplainabilityBaselineConfig :: Lens' DescribeModelExplainabilityJobDefinitionResponse (Maybe ModelExplainabilityBaselineConfig)

The baseline configuration for a model explainability job.

describeModelExplainabilityJobDefinitionResponse_stoppingCondition :: Lens' DescribeModelExplainabilityJobDefinitionResponse (Maybe MonitoringStoppingCondition)

Undocumented member.

describeModelExplainabilityJobDefinitionResponse_httpStatus :: Lens' DescribeModelExplainabilityJobDefinitionResponse Int

The response's http status code.

describeModelExplainabilityJobDefinitionResponse_jobDefinitionArn :: Lens' DescribeModelExplainabilityJobDefinitionResponse Text

The Amazon Resource Name (ARN) of the model explainability job.

describeModelExplainabilityJobDefinitionResponse_jobDefinitionName :: Lens' DescribeModelExplainabilityJobDefinitionResponse Text

The name of the explainability job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

describeModelExplainabilityJobDefinitionResponse_creationTime :: Lens' DescribeModelExplainabilityJobDefinitionResponse UTCTime

The time at which the model explainability job was created.

describeModelExplainabilityJobDefinitionResponse_modelExplainabilityAppSpecification :: Lens' DescribeModelExplainabilityJobDefinitionResponse ModelExplainabilityAppSpecification

Configures the model explainability job to run a specified Docker container image.

describeModelExplainabilityJobDefinitionResponse_modelExplainabilityJobInput :: Lens' DescribeModelExplainabilityJobDefinitionResponse ModelExplainabilityJobInput

Inputs for the model explainability job.

describeModelExplainabilityJobDefinitionResponse_modelExplainabilityJobOutputConfig :: Lens' DescribeModelExplainabilityJobDefinitionResponse MonitoringOutputConfig

Undocumented member.

describeModelExplainabilityJobDefinitionResponse_jobResources :: Lens' DescribeModelExplainabilityJobDefinitionResponse MonitoringResources

Undocumented member.

describeModelExplainabilityJobDefinitionResponse_roleArn :: Lens' DescribeModelExplainabilityJobDefinitionResponse Text

The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that has read permission to the input data location and write permission to the output data location in Amazon S3.