Network.AWS.SageMaker.DescribeModelQualityJobDefinition

Description

Returns a description of a model quality job definition.

Creating a Request

data DescribeModelQualityJobDefinition

See: newDescribeModelQualityJobDefinition smart constructor.

Constructors

DescribeModelQualityJobDefinition'

Fields

jobDefinitionName :: Text

The name of the model quality job. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

newDescribeModelQualityJobDefinition

Arguments

:: Text
$sel:jobDefinitionName:DescribeModelQualityJobDefinition'

-> DescribeModelQualityJobDefinition

Create a value of DescribeModelQualityJobDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobDefinitionName:DescribeModelQualityJobDefinition', describeModelQualityJobDefinition_jobDefinitionName - The name of the model quality job. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

Request Lenses

describeModelQualityJobDefinition_jobDefinitionName :: Lens' DescribeModelQualityJobDefinition Text

The name of the model quality job. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

Destructuring the Response

data DescribeModelQualityJobDefinitionResponse

See: newDescribeModelQualityJobDefinitionResponse smart constructor.

Constructors

DescribeModelQualityJobDefinitionResponse'

Fields

networkConfig :: Maybe MonitoringNetworkConfig

Networking options for a model quality job.

modelQualityBaselineConfig :: Maybe ModelQualityBaselineConfig

The baseline configuration for a model quality job.

stoppingCondition :: Maybe MonitoringStoppingCondition

httpStatus :: Int

The response's http status code.

jobDefinitionArn :: Text

The Amazon Resource Name (ARN) of the model quality job.

jobDefinitionName :: Text

The name of the quality job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

creationTime :: POSIX

The time at which the model quality job was created.

modelQualityAppSpecification :: ModelQualityAppSpecification

Configures the model quality job to run a specified Docker container image.

modelQualityJobInput :: ModelQualityJobInput

Inputs for the model quality job.

modelQualityJobOutputConfig :: MonitoringOutputConfig

jobResources :: MonitoringResources

roleArn :: Text

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

newDescribeModelQualityJobDefinitionResponse

Arguments

:: Int
$sel:httpStatus:DescribeModelQualityJobDefinitionResponse'

-> Text
$sel:jobDefinitionArn:DescribeModelQualityJobDefinitionResponse'

-> Text
$sel:jobDefinitionName:DescribeModelQualityJobDefinition'

-> UTCTime
$sel:creationTime:DescribeModelQualityJobDefinitionResponse'

-> ModelQualityAppSpecification
$sel:modelQualityAppSpecification:DescribeModelQualityJobDefinitionResponse'

-> ModelQualityJobInput
$sel:modelQualityJobInput:DescribeModelQualityJobDefinitionResponse'

-> MonitoringOutputConfig
$sel:modelQualityJobOutputConfig:DescribeModelQualityJobDefinitionResponse'

-> MonitoringResources
$sel:jobResources:DescribeModelQualityJobDefinitionResponse'

-> Text
$sel:roleArn:DescribeModelQualityJobDefinitionResponse'

-> DescribeModelQualityJobDefinitionResponse

Create a value of DescribeModelQualityJobDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkConfig:DescribeModelQualityJobDefinitionResponse', describeModelQualityJobDefinitionResponse_networkConfig - Networking options for a model quality job.

$sel:modelQualityBaselineConfig:DescribeModelQualityJobDefinitionResponse', describeModelQualityJobDefinitionResponse_modelQualityBaselineConfig - The baseline configuration for a model quality job.

$sel:stoppingCondition:DescribeModelQualityJobDefinitionResponse', describeModelQualityJobDefinitionResponse_stoppingCondition - Undocumented member.

$sel:httpStatus:DescribeModelQualityJobDefinitionResponse', describeModelQualityJobDefinitionResponse_httpStatus - The response's http status code.

$sel:jobDefinitionArn:DescribeModelQualityJobDefinitionResponse', describeModelQualityJobDefinitionResponse_jobDefinitionArn - The Amazon Resource Name (ARN) of the model quality job.

$sel:jobDefinitionName:DescribeModelQualityJobDefinition', describeModelQualityJobDefinitionResponse_jobDefinitionName - The name of the quality job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

$sel:creationTime:DescribeModelQualityJobDefinitionResponse', describeModelQualityJobDefinitionResponse_creationTime - The time at which the model quality job was created.

$sel:modelQualityAppSpecification:DescribeModelQualityJobDefinitionResponse', describeModelQualityJobDefinitionResponse_modelQualityAppSpecification - Configures the model quality job to run a specified Docker container image.

$sel:modelQualityJobInput:DescribeModelQualityJobDefinitionResponse', describeModelQualityJobDefinitionResponse_modelQualityJobInput - Inputs for the model quality job.

$sel:modelQualityJobOutputConfig:DescribeModelQualityJobDefinitionResponse', describeModelQualityJobDefinitionResponse_modelQualityJobOutputConfig - Undocumented member.

$sel:jobResources:DescribeModelQualityJobDefinitionResponse', describeModelQualityJobDefinitionResponse_jobResources - Undocumented member.

$sel:roleArn:DescribeModelQualityJobDefinitionResponse', describeModelQualityJobDefinitionResponse_roleArn - The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

Response Lenses

describeModelQualityJobDefinitionResponse_networkConfig :: Lens' DescribeModelQualityJobDefinitionResponse (Maybe MonitoringNetworkConfig)

Networking options for a model quality job.

describeModelQualityJobDefinitionResponse_modelQualityBaselineConfig :: Lens' DescribeModelQualityJobDefinitionResponse (Maybe ModelQualityBaselineConfig)

The baseline configuration for a model quality job.

describeModelQualityJobDefinitionResponse_stoppingCondition :: Lens' DescribeModelQualityJobDefinitionResponse (Maybe MonitoringStoppingCondition)

Undocumented member.

describeModelQualityJobDefinitionResponse_httpStatus :: Lens' DescribeModelQualityJobDefinitionResponse Int

The response's http status code.

describeModelQualityJobDefinitionResponse_jobDefinitionArn :: Lens' DescribeModelQualityJobDefinitionResponse Text

The Amazon Resource Name (ARN) of the model quality job.

describeModelQualityJobDefinitionResponse_jobDefinitionName :: Lens' DescribeModelQualityJobDefinitionResponse Text

The name of the quality job definition. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

describeModelQualityJobDefinitionResponse_creationTime :: Lens' DescribeModelQualityJobDefinitionResponse UTCTime

The time at which the model quality job was created.

describeModelQualityJobDefinitionResponse_modelQualityAppSpecification :: Lens' DescribeModelQualityJobDefinitionResponse ModelQualityAppSpecification

Configures the model quality job to run a specified Docker container image.

describeModelQualityJobDefinitionResponse_modelQualityJobInput :: Lens' DescribeModelQualityJobDefinitionResponse ModelQualityJobInput

Inputs for the model quality job.

describeModelQualityJobDefinitionResponse_modelQualityJobOutputConfig :: Lens' DescribeModelQualityJobDefinitionResponse MonitoringOutputConfig

Undocumented member.

describeModelQualityJobDefinitionResponse_jobResources :: Lens' DescribeModelQualityJobDefinitionResponse MonitoringResources

Undocumented member.

describeModelQualityJobDefinitionResponse_roleArn :: Lens' DescribeModelQualityJobDefinitionResponse Text

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.