Network.AWS.SageMaker.DescribeMonitoringSchedule

Description

Describes the schedule for a monitoring job.

Creating a Request

data DescribeMonitoringSchedule

See: newDescribeMonitoringSchedule smart constructor.

Constructors

DescribeMonitoringSchedule'

Fields

monitoringScheduleName :: Text

Name of a previously created monitoring schedule.

newDescribeMonitoringSchedule

Arguments

:: Text
$sel:monitoringScheduleName:DescribeMonitoringSchedule'

-> DescribeMonitoringSchedule

Create a value of DescribeMonitoringSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:monitoringScheduleName:DescribeMonitoringSchedule', describeMonitoringSchedule_monitoringScheduleName - Name of a previously created monitoring schedule.

Request Lenses

describeMonitoringSchedule_monitoringScheduleName :: Lens' DescribeMonitoringSchedule Text

Name of a previously created monitoring schedule.

Destructuring the Response

data DescribeMonitoringScheduleResponse

See: newDescribeMonitoringScheduleResponse smart constructor.

Constructors

DescribeMonitoringScheduleResponse'

Fields

endpointName :: Maybe Text

The name of the endpoint for the monitoring job.

monitoringType :: Maybe MonitoringType

The type of the monitoring job that this schedule runs. This is one of the following values.

DATA_QUALITY - The schedule is for a data quality monitoring job.

MODEL_QUALITY - The schedule is for a model quality monitoring job.

MODEL_BIAS - The schedule is for a bias monitoring job.

MODEL_EXPLAINABILITY - The schedule is for an explainability monitoring job.

failureReason :: Maybe Text

A string, up to one KB in size, that contains the reason a monitoring job failed, if it failed.

lastMonitoringExecutionSummary :: Maybe MonitoringExecutionSummary

Describes metadata on the last execution to run, if there was one.

httpStatus :: Int

The response's http status code.

monitoringScheduleArn :: Text

The Amazon Resource Name (ARN) of the monitoring schedule.

monitoringScheduleName :: Text

Name of the monitoring schedule.

monitoringScheduleStatus :: ScheduleStatus

The status of an monitoring job.

creationTime :: POSIX

The time at which the monitoring job was created.

lastModifiedTime :: POSIX

The time at which the monitoring job was last modified.

monitoringScheduleConfig :: MonitoringScheduleConfig

The configuration object that specifies the monitoring schedule and defines the monitoring job.

newDescribeMonitoringScheduleResponse

Arguments

:: Int
$sel:httpStatus:DescribeMonitoringScheduleResponse'

-> Text
$sel:monitoringScheduleArn:DescribeMonitoringScheduleResponse'

-> Text
$sel:monitoringScheduleName:DescribeMonitoringSchedule'

-> ScheduleStatus
$sel:monitoringScheduleStatus:DescribeMonitoringScheduleResponse'

-> UTCTime
$sel:creationTime:DescribeMonitoringScheduleResponse'

-> UTCTime
$sel:lastModifiedTime:DescribeMonitoringScheduleResponse'

-> MonitoringScheduleConfig
$sel:monitoringScheduleConfig:DescribeMonitoringScheduleResponse'

-> DescribeMonitoringScheduleResponse

Create a value of DescribeMonitoringScheduleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpointName:DescribeMonitoringScheduleResponse', describeMonitoringScheduleResponse_endpointName - The name of the endpoint for the monitoring job.

$sel:monitoringType:DescribeMonitoringScheduleResponse', describeMonitoringScheduleResponse_monitoringType - The type of the monitoring job that this schedule runs. This is one of the following values.

DATA_QUALITY - The schedule is for a data quality monitoring job.

MODEL_QUALITY - The schedule is for a model quality monitoring job.

MODEL_BIAS - The schedule is for a bias monitoring job.

MODEL_EXPLAINABILITY - The schedule is for an explainability monitoring job.

$sel:failureReason:DescribeMonitoringScheduleResponse', describeMonitoringScheduleResponse_failureReason - A string, up to one KB in size, that contains the reason a monitoring job failed, if it failed.

$sel:lastMonitoringExecutionSummary:DescribeMonitoringScheduleResponse', describeMonitoringScheduleResponse_lastMonitoringExecutionSummary - Describes metadata on the last execution to run, if there was one.

$sel:httpStatus:DescribeMonitoringScheduleResponse', describeMonitoringScheduleResponse_httpStatus - The response's http status code.

$sel:monitoringScheduleArn:DescribeMonitoringScheduleResponse', describeMonitoringScheduleResponse_monitoringScheduleArn - The Amazon Resource Name (ARN) of the monitoring schedule.

$sel:monitoringScheduleName:DescribeMonitoringSchedule', describeMonitoringScheduleResponse_monitoringScheduleName - Name of the monitoring schedule.

$sel:monitoringScheduleStatus:DescribeMonitoringScheduleResponse', describeMonitoringScheduleResponse_monitoringScheduleStatus - The status of an monitoring job.

$sel:creationTime:DescribeMonitoringScheduleResponse', describeMonitoringScheduleResponse_creationTime - The time at which the monitoring job was created.

$sel:lastModifiedTime:DescribeMonitoringScheduleResponse', describeMonitoringScheduleResponse_lastModifiedTime - The time at which the monitoring job was last modified.

$sel:monitoringScheduleConfig:DescribeMonitoringScheduleResponse', describeMonitoringScheduleResponse_monitoringScheduleConfig - The configuration object that specifies the monitoring schedule and defines the monitoring job.

Response Lenses

describeMonitoringScheduleResponse_endpointName :: Lens' DescribeMonitoringScheduleResponse (Maybe Text)

The name of the endpoint for the monitoring job.

describeMonitoringScheduleResponse_monitoringType :: Lens' DescribeMonitoringScheduleResponse (Maybe MonitoringType)

The type of the monitoring job that this schedule runs. This is one of the following values.

DATA_QUALITY - The schedule is for a data quality monitoring job.

MODEL_QUALITY - The schedule is for a model quality monitoring job.

MODEL_BIAS - The schedule is for a bias monitoring job.

MODEL_EXPLAINABILITY - The schedule is for an explainability monitoring job.

describeMonitoringScheduleResponse_failureReason :: Lens' DescribeMonitoringScheduleResponse (Maybe Text)

A string, up to one KB in size, that contains the reason a monitoring job failed, if it failed.

describeMonitoringScheduleResponse_lastMonitoringExecutionSummary :: Lens' DescribeMonitoringScheduleResponse (Maybe MonitoringExecutionSummary)

Describes metadata on the last execution to run, if there was one.

describeMonitoringScheduleResponse_httpStatus :: Lens' DescribeMonitoringScheduleResponse Int

The response's http status code.

describeMonitoringScheduleResponse_monitoringScheduleArn :: Lens' DescribeMonitoringScheduleResponse Text

The Amazon Resource Name (ARN) of the monitoring schedule.

describeMonitoringScheduleResponse_monitoringScheduleName :: Lens' DescribeMonitoringScheduleResponse Text

Name of the monitoring schedule.

describeMonitoringScheduleResponse_monitoringScheduleStatus :: Lens' DescribeMonitoringScheduleResponse ScheduleStatus

The status of an monitoring job.

describeMonitoringScheduleResponse_creationTime :: Lens' DescribeMonitoringScheduleResponse UTCTime

The time at which the monitoring job was created.

describeMonitoringScheduleResponse_lastModifiedTime :: Lens' DescribeMonitoringScheduleResponse UTCTime

The time at which the monitoring job was last modified.

describeMonitoringScheduleResponse_monitoringScheduleConfig :: Lens' DescribeMonitoringScheduleResponse MonitoringScheduleConfig

The configuration object that specifies the monitoring schedule and defines the monitoring job.