Network.AWS.SageMaker.DescribeNotebookInstance

Description

Returns information about a notebook instance.

Creating a Request

data DescribeNotebookInstance

See: newDescribeNotebookInstance smart constructor.

Constructors

DescribeNotebookInstance'

Fields

notebookInstanceName :: Text

The name of the notebook instance that you want information about.

newDescribeNotebookInstance

Arguments

:: Text
$sel:notebookInstanceName:DescribeNotebookInstance'

-> DescribeNotebookInstance

Create a value of DescribeNotebookInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notebookInstanceName:DescribeNotebookInstance', describeNotebookInstance_notebookInstanceName - The name of the notebook instance that you want information about.

Request Lenses

describeNotebookInstance_notebookInstanceName :: Lens' DescribeNotebookInstance Text

The name of the notebook instance that you want information about.

Destructuring the Response

data DescribeNotebookInstanceResponse

See: newDescribeNotebookInstanceResponse smart constructor.

Constructors

DescribeNotebookInstanceResponse'

Fields

notebookInstanceName :: Maybe Text

The name of the Amazon SageMaker notebook instance.

creationTime :: Maybe POSIX

A timestamp. Use this parameter to return the time when the notebook instance was created

platformIdentifier :: Maybe Text

The platform identifier of the notebook instance runtime environment.

defaultCodeRepository :: Maybe Text

The Git repository associated with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in Amazon Web Services CodeCommit or in any other Git repository. When you open a notebook instance, it opens in the directory that contains this repository. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

acceleratorTypes :: Maybe [NotebookInstanceAcceleratorType]

A list of the Elastic Inference (EI) instance types associated with this notebook instance. Currently only one EI instance type can be associated with a notebook instance. For more information, see Using Elastic Inference in Amazon SageMaker.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role associated with the instance.

instanceType :: Maybe InstanceType

The type of ML compute instance running on the notebook instance.

notebookInstanceLifecycleConfigName :: Maybe Text

Returns the name of a notebook instance lifecycle configuration.

For information about notebook instance lifestyle configurations, see Step 2.1: (Optional) Customize a Notebook Instance

additionalCodeRepositories :: Maybe [Text]

An array of up to three Git repositories associated with the notebook instance. These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in Amazon Web Services CodeCommit or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

securityGroups :: Maybe [Text]

The IDs of the VPC security groups.

kmsKeyId :: Maybe Text

The Amazon Web Services KMS key ID Amazon SageMaker uses to encrypt data when storing it on the ML storage volume attached to the instance.

volumeSizeInGB :: Maybe Natural

The size, in GB, of the ML storage volume attached to the notebook instance.

notebookInstanceStatus :: Maybe NotebookInstanceStatus

The status of the notebook instance.

failureReason :: Maybe Text

If status is Failed, the reason it failed.

lastModifiedTime :: Maybe POSIX

A timestamp. Use this parameter to retrieve the time when the notebook instance was last modified.

subnetId :: Maybe Text

The ID of the VPC subnet.

networkInterfaceId :: Maybe Text

The network interface IDs that Amazon SageMaker created at the time of creating the instance.

url :: Maybe Text

The URL that you use to connect to the Jupyter notebook that is running in your notebook instance.

notebookInstanceArn :: Maybe Text

The Amazon Resource Name (ARN) of the notebook instance.

rootAccess :: Maybe RootAccess

Whether root access is enabled or disabled for users of the notebook instance.

Lifecycle configurations need root access to be able to set up a notebook instance. Because of this, lifecycle configurations associated with a notebook instance always run with root access even if you disable root access for users.

directInternetAccess :: Maybe DirectInternetAccess

Describes whether Amazon SageMaker provides internet access to the notebook instance. If this value is set to Disabled, the notebook instance does not have internet access, and cannot connect to Amazon SageMaker training and endpoint services.

For more information, see Notebook Instances Are Internet-Enabled by Default.

httpStatus :: Int

The response's http status code.

newDescribeNotebookInstanceResponse

Arguments

:: Int
$sel:httpStatus:DescribeNotebookInstanceResponse'

-> DescribeNotebookInstanceResponse

Create a value of DescribeNotebookInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notebookInstanceName:DescribeNotebookInstance', describeNotebookInstanceResponse_notebookInstanceName - The name of the Amazon SageMaker notebook instance.

$sel:creationTime:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_creationTime - A timestamp. Use this parameter to return the time when the notebook instance was created

$sel:platformIdentifier:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_platformIdentifier - The platform identifier of the notebook instance runtime environment.

$sel:defaultCodeRepository:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_defaultCodeRepository - The Git repository associated with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in Amazon Web Services CodeCommit or in any other Git repository. When you open a notebook instance, it opens in the directory that contains this repository. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

$sel:acceleratorTypes:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_acceleratorTypes - A list of the Elastic Inference (EI) instance types associated with this notebook instance. Currently only one EI instance type can be associated with a notebook instance. For more information, see Using Elastic Inference in Amazon SageMaker.

$sel:roleArn:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_roleArn - The Amazon Resource Name (ARN) of the IAM role associated with the instance.

$sel:instanceType:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_instanceType - The type of ML compute instance running on the notebook instance.

$sel:notebookInstanceLifecycleConfigName:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_notebookInstanceLifecycleConfigName - Returns the name of a notebook instance lifecycle configuration.

For information about notebook instance lifestyle configurations, see Step 2.1: (Optional) Customize a Notebook Instance

$sel:additionalCodeRepositories:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_additionalCodeRepositories - An array of up to three Git repositories associated with the notebook instance. These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in Amazon Web Services CodeCommit or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

$sel:securityGroups:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_securityGroups - The IDs of the VPC security groups.

$sel:kmsKeyId:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_kmsKeyId - The Amazon Web Services KMS key ID Amazon SageMaker uses to encrypt data when storing it on the ML storage volume attached to the instance.

$sel:volumeSizeInGB:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_volumeSizeInGB - The size, in GB, of the ML storage volume attached to the notebook instance.

$sel:notebookInstanceStatus:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_notebookInstanceStatus - The status of the notebook instance.

$sel:failureReason:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_failureReason - If status is Failed, the reason it failed.

$sel:lastModifiedTime:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_lastModifiedTime - A timestamp. Use this parameter to retrieve the time when the notebook instance was last modified.

$sel:subnetId:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_subnetId - The ID of the VPC subnet.

$sel:networkInterfaceId:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_networkInterfaceId - The network interface IDs that Amazon SageMaker created at the time of creating the instance.

$sel:url:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_url - The URL that you use to connect to the Jupyter notebook that is running in your notebook instance.

$sel:notebookInstanceArn:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_notebookInstanceArn - The Amazon Resource Name (ARN) of the notebook instance.

$sel:rootAccess:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_rootAccess - Whether root access is enabled or disabled for users of the notebook instance.

Lifecycle configurations need root access to be able to set up a notebook instance. Because of this, lifecycle configurations associated with a notebook instance always run with root access even if you disable root access for users.

$sel:directInternetAccess:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_directInternetAccess - Describes whether Amazon SageMaker provides internet access to the notebook instance. If this value is set to Disabled, the notebook instance does not have internet access, and cannot connect to Amazon SageMaker training and endpoint services.

For more information, see Notebook Instances Are Internet-Enabled by Default.

$sel:httpStatus:DescribeNotebookInstanceResponse', describeNotebookInstanceResponse_httpStatus - The response's http status code.

Response Lenses

describeNotebookInstanceResponse_notebookInstanceName :: Lens' DescribeNotebookInstanceResponse (Maybe Text)

The name of the Amazon SageMaker notebook instance.

describeNotebookInstanceResponse_creationTime :: Lens' DescribeNotebookInstanceResponse (Maybe UTCTime)

A timestamp. Use this parameter to return the time when the notebook instance was created

describeNotebookInstanceResponse_platformIdentifier :: Lens' DescribeNotebookInstanceResponse (Maybe Text)

The platform identifier of the notebook instance runtime environment.

describeNotebookInstanceResponse_defaultCodeRepository :: Lens' DescribeNotebookInstanceResponse (Maybe Text)

The Git repository associated with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in Amazon Web Services CodeCommit or in any other Git repository. When you open a notebook instance, it opens in the directory that contains this repository. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

describeNotebookInstanceResponse_acceleratorTypes :: Lens' DescribeNotebookInstanceResponse (Maybe [NotebookInstanceAcceleratorType])

A list of the Elastic Inference (EI) instance types associated with this notebook instance. Currently only one EI instance type can be associated with a notebook instance. For more information, see Using Elastic Inference in Amazon SageMaker.

describeNotebookInstanceResponse_roleArn :: Lens' DescribeNotebookInstanceResponse (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role associated with the instance.

describeNotebookInstanceResponse_instanceType :: Lens' DescribeNotebookInstanceResponse (Maybe InstanceType)

The type of ML compute instance running on the notebook instance.

describeNotebookInstanceResponse_notebookInstanceLifecycleConfigName :: Lens' DescribeNotebookInstanceResponse (Maybe Text)

Returns the name of a notebook instance lifecycle configuration.

For information about notebook instance lifestyle configurations, see Step 2.1: (Optional) Customize a Notebook Instance

describeNotebookInstanceResponse_additionalCodeRepositories :: Lens' DescribeNotebookInstanceResponse (Maybe [Text])

An array of up to three Git repositories associated with the notebook instance. These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in Amazon Web Services CodeCommit or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

describeNotebookInstanceResponse_securityGroups :: Lens' DescribeNotebookInstanceResponse (Maybe [Text])

The IDs of the VPC security groups.

describeNotebookInstanceResponse_kmsKeyId :: Lens' DescribeNotebookInstanceResponse (Maybe Text)

The Amazon Web Services KMS key ID Amazon SageMaker uses to encrypt data when storing it on the ML storage volume attached to the instance.

describeNotebookInstanceResponse_volumeSizeInGB :: Lens' DescribeNotebookInstanceResponse (Maybe Natural)

The size, in GB, of the ML storage volume attached to the notebook instance.

describeNotebookInstanceResponse_notebookInstanceStatus :: Lens' DescribeNotebookInstanceResponse (Maybe NotebookInstanceStatus)

The status of the notebook instance.

describeNotebookInstanceResponse_failureReason :: Lens' DescribeNotebookInstanceResponse (Maybe Text)

If status is Failed, the reason it failed.

describeNotebookInstanceResponse_lastModifiedTime :: Lens' DescribeNotebookInstanceResponse (Maybe UTCTime)

A timestamp. Use this parameter to retrieve the time when the notebook instance was last modified.

describeNotebookInstanceResponse_subnetId :: Lens' DescribeNotebookInstanceResponse (Maybe Text)

The ID of the VPC subnet.

describeNotebookInstanceResponse_networkInterfaceId :: Lens' DescribeNotebookInstanceResponse (Maybe Text)

The network interface IDs that Amazon SageMaker created at the time of creating the instance.

describeNotebookInstanceResponse_url :: Lens' DescribeNotebookInstanceResponse (Maybe Text)

The URL that you use to connect to the Jupyter notebook that is running in your notebook instance.

describeNotebookInstanceResponse_notebookInstanceArn :: Lens' DescribeNotebookInstanceResponse (Maybe Text)

The Amazon Resource Name (ARN) of the notebook instance.

describeNotebookInstanceResponse_rootAccess :: Lens' DescribeNotebookInstanceResponse (Maybe RootAccess)

Whether root access is enabled or disabled for users of the notebook instance.

Lifecycle configurations need root access to be able to set up a notebook instance. Because of this, lifecycle configurations associated with a notebook instance always run with root access even if you disable root access for users.

describeNotebookInstanceResponse_directInternetAccess :: Lens' DescribeNotebookInstanceResponse (Maybe DirectInternetAccess)

Describes whether Amazon SageMaker provides internet access to the notebook instance. If this value is set to Disabled, the notebook instance does not have internet access, and cannot connect to Amazon SageMaker training and endpoint services.

For more information, see Notebook Instances Are Internet-Enabled by Default.

describeNotebookInstanceResponse_httpStatus :: Lens' DescribeNotebookInstanceResponse Int

The response's http status code.