Network.AWS.SageMaker.DescribeNotebookInstanceLifecycleConfig

Description

Returns a description of a notebook instance lifecycle configuration.

For information about notebook instance lifestyle configurations, see Step 2.1: (Optional) Customize a Notebook Instance.

Creating a Request

data DescribeNotebookInstanceLifecycleConfig

See: newDescribeNotebookInstanceLifecycleConfig smart constructor.

Constructors

DescribeNotebookInstanceLifecycleConfig'

Fields

notebookInstanceLifecycleConfigName :: Text

The name of the lifecycle configuration to describe.

newDescribeNotebookInstanceLifecycleConfig

Arguments

:: Text
$sel:notebookInstanceLifecycleConfigName:DescribeNotebookInstanceLifecycleConfig'

-> DescribeNotebookInstanceLifecycleConfig

Create a value of DescribeNotebookInstanceLifecycleConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notebookInstanceLifecycleConfigName:DescribeNotebookInstanceLifecycleConfig', describeNotebookInstanceLifecycleConfig_notebookInstanceLifecycleConfigName - The name of the lifecycle configuration to describe.

Request Lenses

describeNotebookInstanceLifecycleConfig_notebookInstanceLifecycleConfigName :: Lens' DescribeNotebookInstanceLifecycleConfig Text

The name of the lifecycle configuration to describe.

Destructuring the Response

data DescribeNotebookInstanceLifecycleConfigResponse

See: newDescribeNotebookInstanceLifecycleConfigResponse smart constructor.

Constructors

DescribeNotebookInstanceLifecycleConfigResponse'

Fields

creationTime :: Maybe POSIX

A timestamp that tells when the lifecycle configuration was created.

notebookInstanceLifecycleConfigArn :: Maybe Text

The Amazon Resource Name (ARN) of the lifecycle configuration.

notebookInstanceLifecycleConfigName :: Maybe Text

The name of the lifecycle configuration.

onStart :: Maybe [NotebookInstanceLifecycleHook]

The shell script that runs every time you start a notebook instance, including when you create the notebook instance.

lastModifiedTime :: Maybe POSIX

A timestamp that tells when the lifecycle configuration was last modified.

onCreate :: Maybe [NotebookInstanceLifecycleHook]

The shell script that runs only once, when you create a notebook instance.

httpStatus :: Int

The response's http status code.

newDescribeNotebookInstanceLifecycleConfigResponse

Arguments

:: Int
$sel:httpStatus:DescribeNotebookInstanceLifecycleConfigResponse'

-> DescribeNotebookInstanceLifecycleConfigResponse

Create a value of DescribeNotebookInstanceLifecycleConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeNotebookInstanceLifecycleConfigResponse', describeNotebookInstanceLifecycleConfigResponse_creationTime - A timestamp that tells when the lifecycle configuration was created.

$sel:notebookInstanceLifecycleConfigArn:DescribeNotebookInstanceLifecycleConfigResponse', describeNotebookInstanceLifecycleConfigResponse_notebookInstanceLifecycleConfigArn - The Amazon Resource Name (ARN) of the lifecycle configuration.

$sel:notebookInstanceLifecycleConfigName:DescribeNotebookInstanceLifecycleConfig', describeNotebookInstanceLifecycleConfigResponse_notebookInstanceLifecycleConfigName - The name of the lifecycle configuration.

$sel:onStart:DescribeNotebookInstanceLifecycleConfigResponse', describeNotebookInstanceLifecycleConfigResponse_onStart - The shell script that runs every time you start a notebook instance, including when you create the notebook instance.

$sel:lastModifiedTime:DescribeNotebookInstanceLifecycleConfigResponse', describeNotebookInstanceLifecycleConfigResponse_lastModifiedTime - A timestamp that tells when the lifecycle configuration was last modified.

$sel:onCreate:DescribeNotebookInstanceLifecycleConfigResponse', describeNotebookInstanceLifecycleConfigResponse_onCreate - The shell script that runs only once, when you create a notebook instance.

$sel:httpStatus:DescribeNotebookInstanceLifecycleConfigResponse', describeNotebookInstanceLifecycleConfigResponse_httpStatus - The response's http status code.

Response Lenses

describeNotebookInstanceLifecycleConfigResponse_creationTime :: Lens' DescribeNotebookInstanceLifecycleConfigResponse (Maybe UTCTime)

A timestamp that tells when the lifecycle configuration was created.

describeNotebookInstanceLifecycleConfigResponse_notebookInstanceLifecycleConfigArn :: Lens' DescribeNotebookInstanceLifecycleConfigResponse (Maybe Text)

The Amazon Resource Name (ARN) of the lifecycle configuration.

describeNotebookInstanceLifecycleConfigResponse_notebookInstanceLifecycleConfigName :: Lens' DescribeNotebookInstanceLifecycleConfigResponse (Maybe Text)

The name of the lifecycle configuration.

describeNotebookInstanceLifecycleConfigResponse_onStart :: Lens' DescribeNotebookInstanceLifecycleConfigResponse (Maybe [NotebookInstanceLifecycleHook])

The shell script that runs every time you start a notebook instance, including when you create the notebook instance.

describeNotebookInstanceLifecycleConfigResponse_lastModifiedTime :: Lens' DescribeNotebookInstanceLifecycleConfigResponse (Maybe UTCTime)

A timestamp that tells when the lifecycle configuration was last modified.

describeNotebookInstanceLifecycleConfigResponse_onCreate :: Lens' DescribeNotebookInstanceLifecycleConfigResponse (Maybe [NotebookInstanceLifecycleHook])

The shell script that runs only once, when you create a notebook instance.

describeNotebookInstanceLifecycleConfigResponse_httpStatus :: Lens' DescribeNotebookInstanceLifecycleConfigResponse Int

The response's http status code.