Network.AWS.SageMaker.DescribeProcessingJob

Description

Returns a description of a processing job.

Creating a Request

data DescribeProcessingJob

See: newDescribeProcessingJob smart constructor.

Constructors

DescribeProcessingJob'

Fields

processingJobName :: Text

The name of the processing job. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

newDescribeProcessingJob

Arguments

:: Text
$sel:processingJobName:DescribeProcessingJob'

-> DescribeProcessingJob

Create a value of DescribeProcessingJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:processingJobName:DescribeProcessingJob', describeProcessingJob_processingJobName - The name of the processing job. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

Request Lenses

describeProcessingJob_processingJobName :: Lens' DescribeProcessingJob Text

The name of the processing job. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

Destructuring the Response

data DescribeProcessingJobResponse

See: newDescribeProcessingJobResponse smart constructor.

Constructors

DescribeProcessingJobResponse'

Fields

networkConfig :: Maybe NetworkConfig

Networking options for a processing job.

processingEndTime :: Maybe POSIX

The time at which the processing job completed.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

processingOutputConfig :: Maybe ProcessingOutputConfig

Output configuration for the processing job.

experimentConfig :: Maybe ExperimentConfig

The configuration information used to create an experiment.

exitMessage :: Maybe Text

An optional string, up to one KB in size, that contains metadata from the processing container when the processing job exits.

environment :: Maybe (HashMap Text Text)

The environment variables set in the Docker container.

autoMLJobArn :: Maybe Text

The ARN of an AutoML job associated with this processing job.

monitoringScheduleArn :: Maybe Text

The ARN of a monitoring schedule for an endpoint associated with this processing job.

failureReason :: Maybe Text

A string, up to one KB in size, that contains the reason a processing job failed, if it failed.

lastModifiedTime :: Maybe POSIX

The time at which the processing job was last modified.

processingInputs :: Maybe [ProcessingInput]

The inputs for a processing job.

processingStartTime :: Maybe POSIX

The time at which the processing job started.

stoppingCondition :: Maybe ProcessingStoppingCondition

The time limit for how long the processing job is allowed to run.

trainingJobArn :: Maybe Text

The ARN of a training job associated with this processing job.

httpStatus :: Int

The response's http status code.

processingJobName :: Text

The name of the processing job. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

processingResources :: ProcessingResources

Identifies the resources, ML compute instances, and ML storage volumes to deploy for a processing job. In distributed training, you specify more than one instance.

appSpecification :: AppSpecification

Configures the processing job to run a specified container image.

processingJobArn :: Text

The Amazon Resource Name (ARN) of the processing job.

processingJobStatus :: ProcessingJobStatus

Provides the status of a processing job.

creationTime :: POSIX

The time at which the processing job was created.

newDescribeProcessingJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeProcessingJobResponse'

-> Text
$sel:processingJobName:DescribeProcessingJob'

-> ProcessingResources
$sel:processingResources:DescribeProcessingJobResponse'

-> AppSpecification
$sel:appSpecification:DescribeProcessingJobResponse'

-> Text
$sel:processingJobArn:DescribeProcessingJobResponse'

-> ProcessingJobStatus
$sel:processingJobStatus:DescribeProcessingJobResponse'

-> UTCTime
$sel:creationTime:DescribeProcessingJobResponse'

-> DescribeProcessingJobResponse

Create a value of DescribeProcessingJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkConfig:DescribeProcessingJobResponse', describeProcessingJobResponse_networkConfig - Networking options for a processing job.

$sel:processingEndTime:DescribeProcessingJobResponse', describeProcessingJobResponse_processingEndTime - The time at which the processing job completed.

$sel:roleArn:DescribeProcessingJobResponse', describeProcessingJobResponse_roleArn - The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

$sel:processingOutputConfig:DescribeProcessingJobResponse', describeProcessingJobResponse_processingOutputConfig - Output configuration for the processing job.

$sel:experimentConfig:DescribeProcessingJobResponse', describeProcessingJobResponse_experimentConfig - The configuration information used to create an experiment.

$sel:exitMessage:DescribeProcessingJobResponse', describeProcessingJobResponse_exitMessage - An optional string, up to one KB in size, that contains metadata from the processing container when the processing job exits.

$sel:environment:DescribeProcessingJobResponse', describeProcessingJobResponse_environment - The environment variables set in the Docker container.

$sel:autoMLJobArn:DescribeProcessingJobResponse', describeProcessingJobResponse_autoMLJobArn - The ARN of an AutoML job associated with this processing job.

$sel:monitoringScheduleArn:DescribeProcessingJobResponse', describeProcessingJobResponse_monitoringScheduleArn - The ARN of a monitoring schedule for an endpoint associated with this processing job.

$sel:failureReason:DescribeProcessingJobResponse', describeProcessingJobResponse_failureReason - A string, up to one KB in size, that contains the reason a processing job failed, if it failed.

$sel:lastModifiedTime:DescribeProcessingJobResponse', describeProcessingJobResponse_lastModifiedTime - The time at which the processing job was last modified.

$sel:processingInputs:DescribeProcessingJobResponse', describeProcessingJobResponse_processingInputs - The inputs for a processing job.

$sel:processingStartTime:DescribeProcessingJobResponse', describeProcessingJobResponse_processingStartTime - The time at which the processing job started.

$sel:stoppingCondition:DescribeProcessingJobResponse', describeProcessingJobResponse_stoppingCondition - The time limit for how long the processing job is allowed to run.

$sel:trainingJobArn:DescribeProcessingJobResponse', describeProcessingJobResponse_trainingJobArn - The ARN of a training job associated with this processing job.

$sel:httpStatus:DescribeProcessingJobResponse', describeProcessingJobResponse_httpStatus - The response's http status code.

$sel:processingJobName:DescribeProcessingJob', describeProcessingJobResponse_processingJobName - The name of the processing job. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

$sel:processingResources:DescribeProcessingJobResponse', describeProcessingJobResponse_processingResources - Identifies the resources, ML compute instances, and ML storage volumes to deploy for a processing job. In distributed training, you specify more than one instance.

$sel:appSpecification:DescribeProcessingJobResponse', describeProcessingJobResponse_appSpecification - Configures the processing job to run a specified container image.

$sel:processingJobArn:DescribeProcessingJobResponse', describeProcessingJobResponse_processingJobArn - The Amazon Resource Name (ARN) of the processing job.

$sel:processingJobStatus:DescribeProcessingJobResponse', describeProcessingJobResponse_processingJobStatus - Provides the status of a processing job.

$sel:creationTime:DescribeProcessingJobResponse', describeProcessingJobResponse_creationTime - The time at which the processing job was created.

Response Lenses

describeProcessingJobResponse_networkConfig :: Lens' DescribeProcessingJobResponse (Maybe NetworkConfig)

Networking options for a processing job.

describeProcessingJobResponse_processingEndTime :: Lens' DescribeProcessingJobResponse (Maybe UTCTime)

The time at which the processing job completed.

describeProcessingJobResponse_roleArn :: Lens' DescribeProcessingJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

describeProcessingJobResponse_processingOutputConfig :: Lens' DescribeProcessingJobResponse (Maybe ProcessingOutputConfig)

Output configuration for the processing job.

describeProcessingJobResponse_experimentConfig :: Lens' DescribeProcessingJobResponse (Maybe ExperimentConfig)

The configuration information used to create an experiment.

describeProcessingJobResponse_exitMessage :: Lens' DescribeProcessingJobResponse (Maybe Text)

An optional string, up to one KB in size, that contains metadata from the processing container when the processing job exits.

describeProcessingJobResponse_environment :: Lens' DescribeProcessingJobResponse (Maybe (HashMap Text Text))

The environment variables set in the Docker container.

describeProcessingJobResponse_autoMLJobArn :: Lens' DescribeProcessingJobResponse (Maybe Text)

The ARN of an AutoML job associated with this processing job.

describeProcessingJobResponse_monitoringScheduleArn :: Lens' DescribeProcessingJobResponse (Maybe Text)

The ARN of a monitoring schedule for an endpoint associated with this processing job.

describeProcessingJobResponse_failureReason :: Lens' DescribeProcessingJobResponse (Maybe Text)

A string, up to one KB in size, that contains the reason a processing job failed, if it failed.

describeProcessingJobResponse_lastModifiedTime :: Lens' DescribeProcessingJobResponse (Maybe UTCTime)

The time at which the processing job was last modified.

describeProcessingJobResponse_processingInputs :: Lens' DescribeProcessingJobResponse (Maybe [ProcessingInput])

The inputs for a processing job.

describeProcessingJobResponse_processingStartTime :: Lens' DescribeProcessingJobResponse (Maybe UTCTime)

The time at which the processing job started.

describeProcessingJobResponse_stoppingCondition :: Lens' DescribeProcessingJobResponse (Maybe ProcessingStoppingCondition)

The time limit for how long the processing job is allowed to run.

describeProcessingJobResponse_trainingJobArn :: Lens' DescribeProcessingJobResponse (Maybe Text)

The ARN of a training job associated with this processing job.

describeProcessingJobResponse_httpStatus :: Lens' DescribeProcessingJobResponse Int

The response's http status code.

describeProcessingJobResponse_processingJobName :: Lens' DescribeProcessingJobResponse Text

The name of the processing job. The name must be unique within an Amazon Web Services Region in the Amazon Web Services account.

describeProcessingJobResponse_processingResources :: Lens' DescribeProcessingJobResponse ProcessingResources

Identifies the resources, ML compute instances, and ML storage volumes to deploy for a processing job. In distributed training, you specify more than one instance.

describeProcessingJobResponse_appSpecification :: Lens' DescribeProcessingJobResponse AppSpecification

Configures the processing job to run a specified container image.

describeProcessingJobResponse_processingJobArn :: Lens' DescribeProcessingJobResponse Text

The Amazon Resource Name (ARN) of the processing job.

describeProcessingJobResponse_processingJobStatus :: Lens' DescribeProcessingJobResponse ProcessingJobStatus

Provides the status of a processing job.

describeProcessingJobResponse_creationTime :: Lens' DescribeProcessingJobResponse UTCTime

The time at which the processing job was created.