Network.AWS.SageMaker.DescribeProject

Description

Describes the details of a project.

Creating a Request

data DescribeProject

See: newDescribeProject smart constructor.

Constructors

DescribeProject'

Fields

projectName :: Text

The name of the project to describe.

newDescribeProject

Arguments

:: Text
$sel:projectName:DescribeProject'

-> DescribeProject

Create a value of DescribeProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectName:DescribeProject', describeProject_projectName - The name of the project to describe.

Request Lenses

describeProject_projectName :: Lens' DescribeProject Text

The name of the project to describe.

Destructuring the Response

data DescribeProjectResponse

See: newDescribeProjectResponse smart constructor.

Constructors

DescribeProjectResponse'

Fields

serviceCatalogProvisionedProductDetails :: Maybe ServiceCatalogProvisionedProductDetails

Information about a provisioned service catalog product.

createdBy :: Maybe UserContext

projectDescription :: Maybe Text

The description of the project.

httpStatus :: Int

The response's http status code.

projectArn :: Text

The Amazon Resource Name (ARN) of the project.

projectName :: Text

The name of the project.

projectId :: Text

The ID of the project.

serviceCatalogProvisioningDetails :: ServiceCatalogProvisioningDetails

Information used to provision a service catalog product. For information, see What is Amazon Web Services Service Catalog.

projectStatus :: ProjectStatus

The status of the project.

creationTime :: POSIX

The time when the project was created.

newDescribeProjectResponse

Arguments

:: Int
$sel:httpStatus:DescribeProjectResponse'

-> Text
$sel:projectArn:DescribeProjectResponse'

-> Text
$sel:projectName:DescribeProject'

-> Text
$sel:projectId:DescribeProjectResponse'

-> ServiceCatalogProvisioningDetails
$sel:serviceCatalogProvisioningDetails:DescribeProjectResponse'

-> ProjectStatus
$sel:projectStatus:DescribeProjectResponse'

-> UTCTime
$sel:creationTime:DescribeProjectResponse'

-> DescribeProjectResponse

Create a value of DescribeProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceCatalogProvisionedProductDetails:DescribeProjectResponse', describeProjectResponse_serviceCatalogProvisionedProductDetails - Information about a provisioned service catalog product.

$sel:createdBy:DescribeProjectResponse', describeProjectResponse_createdBy - Undocumented member.

$sel:projectDescription:DescribeProjectResponse', describeProjectResponse_projectDescription - The description of the project.

$sel:httpStatus:DescribeProjectResponse', describeProjectResponse_httpStatus - The response's http status code.

$sel:projectArn:DescribeProjectResponse', describeProjectResponse_projectArn - The Amazon Resource Name (ARN) of the project.

$sel:projectName:DescribeProject', describeProjectResponse_projectName - The name of the project.

$sel:projectId:DescribeProjectResponse', describeProjectResponse_projectId - The ID of the project.

$sel:serviceCatalogProvisioningDetails:DescribeProjectResponse', describeProjectResponse_serviceCatalogProvisioningDetails - Information used to provision a service catalog product. For information, see What is Amazon Web Services Service Catalog.

$sel:projectStatus:DescribeProjectResponse', describeProjectResponse_projectStatus - The status of the project.

$sel:creationTime:DescribeProjectResponse', describeProjectResponse_creationTime - The time when the project was created.

Response Lenses

describeProjectResponse_serviceCatalogProvisionedProductDetails :: Lens' DescribeProjectResponse (Maybe ServiceCatalogProvisionedProductDetails)

Information about a provisioned service catalog product.

describeProjectResponse_createdBy :: Lens' DescribeProjectResponse (Maybe UserContext)

Undocumented member.

describeProjectResponse_projectDescription :: Lens' DescribeProjectResponse (Maybe Text)

The description of the project.

describeProjectResponse_httpStatus :: Lens' DescribeProjectResponse Int

The response's http status code.

describeProjectResponse_projectArn :: Lens' DescribeProjectResponse Text

The Amazon Resource Name (ARN) of the project.

describeProjectResponse_projectName :: Lens' DescribeProjectResponse Text

The name of the project.

describeProjectResponse_projectId :: Lens' DescribeProjectResponse Text

The ID of the project.

describeProjectResponse_serviceCatalogProvisioningDetails :: Lens' DescribeProjectResponse ServiceCatalogProvisioningDetails

Information used to provision a service catalog product. For information, see What is Amazon Web Services Service Catalog.

describeProjectResponse_projectStatus :: Lens' DescribeProjectResponse ProjectStatus

The status of the project.

describeProjectResponse_creationTime :: Lens' DescribeProjectResponse UTCTime

The time when the project was created.