Network.AWS.SageMaker.DescribeTrainingJob

Description

Returns information about a training job.

Some of the attributes below only appear if the training job successfully starts. If the training job fails, TrainingJobStatus is Failed and, depending on the FailureReason, attributes like TrainingStartTime, TrainingTimeInSeconds, TrainingEndTime, and BillableTimeInSeconds may not be present in the response.

Creating a Request

data DescribeTrainingJob

See: newDescribeTrainingJob smart constructor.

Constructors

DescribeTrainingJob'

Fields

trainingJobName :: Text

The name of the training job.

newDescribeTrainingJob

Arguments

:: Text
$sel:trainingJobName:DescribeTrainingJob'

-> DescribeTrainingJob

Create a value of DescribeTrainingJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trainingJobName:DescribeTrainingJob', describeTrainingJob_trainingJobName - The name of the training job.

Request Lenses

describeTrainingJob_trainingJobName :: Lens' DescribeTrainingJob Text

The name of the training job.

Destructuring the Response

data DescribeTrainingJobResponse

See: newDescribeTrainingJobResponse smart constructor.

Constructors

DescribeTrainingJobResponse'

Fields

vpcConfig :: Maybe VpcConfig

A VpcConfig object that specifies the VPC that this training job has access to. For more information, see Protect Training Jobs by Using an Amazon Virtual Private Cloud.

enableManagedSpotTraining :: Maybe Bool

A Boolean indicating whether managed spot training is enabled (True) or not (False).

labelingJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon SageMaker Ground Truth labeling job that created the transform or training job.

hyperParameters :: Maybe (HashMap Text Text)

Algorithm-specific parameters.

debugRuleConfigurations :: Maybe [DebugRuleConfiguration]

Configuration information for Debugger rules for debugging output tensors.

inputDataConfig :: Maybe (NonEmpty Channel)

An array of Channel objects that describes each data input channel.

trainingTimeInSeconds :: Maybe Natural

The training time in seconds.

roleArn :: Maybe Text

The Amazon Web Services Identity and Access Management (IAM) role configured for the training job.

experimentConfig :: Maybe ExperimentConfig

enableNetworkIsolation :: Maybe Bool

If you want to allow inbound or outbound network calls, except for calls between peers within a training cluster for distributed training, choose True. If you enable network isolation for training jobs that are configured to use a VPC, Amazon SageMaker downloads and uploads customer data and model artifacts through the specified VPC, but the training container does not have network access.

profilerRuleEvaluationStatuses :: Maybe [ProfilerRuleEvaluationStatus]

Evaluation status of Debugger rules for profiling on a training job.

profilerConfig :: Maybe ProfilerConfig

enableInterContainerTrafficEncryption :: Maybe Bool

To encrypt all communications between ML compute instances in distributed training, choose True. Encryption provides greater security for distributed training, but training might take longer. How long it takes depends on the amount of communication between compute instances, especially if you use a deep learning algorithms in distributed training.

checkpointConfig :: Maybe CheckpointConfig

tuningJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the associated hyperparameter tuning job if the training job was launched by a hyperparameter tuning job.

outputDataConfig :: Maybe OutputDataConfig

The S3 path where model artifacts that you configured when creating the job are stored. Amazon SageMaker creates subfolders for model artifacts.

environment :: Maybe (HashMap Text Text)

The environment variables to set in the Docker container.

profilingStatus :: Maybe ProfilingStatus

Profiling status of a training job.

finalMetricDataList :: Maybe [MetricData]

A collection of MetricData objects that specify the names, values, and dates and times that the training algorithm emitted to Amazon CloudWatch.

secondaryStatusTransitions :: Maybe [SecondaryStatusTransition]

A history of all of the secondary statuses that the training job has transitioned through.

autoMLJobArn :: Maybe Text

The Amazon Resource Name (ARN) of an AutoML job.

profilerRuleConfigurations :: Maybe [ProfilerRuleConfiguration]

Configuration information for Debugger rules for profiling system and framework metrics.

failureReason :: Maybe Text

If the training job failed, the reason it failed.

lastModifiedTime :: Maybe POSIX

A timestamp that indicates when the status of the training job was last modified.

tensorBoardOutputConfig :: Maybe TensorBoardOutputConfig

debugRuleEvaluationStatuses :: Maybe [DebugRuleEvaluationStatus]

Evaluation status of Debugger rules for debugging on a training job.

retryStrategy :: Maybe RetryStrategy

The number of times to retry the job when the job fails due to an InternalServerError.

debugHookConfig :: Maybe DebugHookConfig

billableTimeInSeconds :: Maybe Natural

The billable time in seconds. Billable time refers to the absolute wall-clock time.

Multiply BillableTimeInSeconds by the number of instances (InstanceCount) in your training cluster to get the total compute time Amazon SageMaker will bill you if you run distributed training. The formula is as follows: BillableTimeInSeconds * InstanceCount .

You can calculate the savings from using managed spot training using the formula (1 - BillableTimeInSeconds / TrainingTimeInSeconds) * 100. For example, if BillableTimeInSeconds is 100 and TrainingTimeInSeconds is 500, the savings is 80%.

trainingStartTime :: Maybe POSIX

Indicates the time when the training job starts on training instances. You are billed for the time interval between this time and the value of TrainingEndTime. The start time in CloudWatch Logs might be later than this time. The difference is due to the time it takes to download the training data and to the size of the training container.

trainingEndTime :: Maybe POSIX

Indicates the time when the training job ends on training instances. You are billed for the time interval between the value of TrainingStartTime and this time. For successful jobs and stopped jobs, this is the time after model artifacts are uploaded. For failed jobs, this is the time when Amazon SageMaker detects a job failure.

httpStatus :: Int

The response's http status code.

trainingJobName :: Text

Name of the model training job.

trainingJobArn :: Text

The Amazon Resource Name (ARN) of the training job.

modelArtifacts :: ModelArtifacts

Information about the Amazon S3 location that is configured for storing model artifacts.

trainingJobStatus :: TrainingJobStatus

The status of the training job.

Amazon SageMaker provides the following training job statuses:

InProgress - The training is in progress.

Completed - The training job has completed.

Failed - The training job has failed. To see the reason for the failure, see the FailureReason field in the response to a DescribeTrainingJobResponse call.

Stopping - The training job is stopping.

Stopped - The training job has stopped.

For more detailed information, see SecondaryStatus.

secondaryStatus :: SecondaryStatus

Provides detailed information about the state of the training job. For detailed information on the secondary status of the training job, see StatusMessage under SecondaryStatusTransition.

Amazon SageMaker provides primary statuses and secondary statuses that apply to each of them:

InProgress

- Starting - Starting the training job.

Downloading - An optional stage for algorithms that support File training input mode. It indicates that data is being downloaded to the ML storage volumes.

Training - Training is in progress.

Interrupted - The job stopped because the managed spot training instances were interrupted.

Uploading - Training is complete and the model artifacts are being uploaded to the S3 location.

Completed

- Completed - The training job has completed.

Failed

- Failed - The training job has failed. The reason for the failure is returned in the FailureReason field of DescribeTrainingJobResponse.

Stopped

- MaxRuntimeExceeded - The job stopped because it exceeded the maximum allowed runtime.

MaxWaitTimeExceeded - The job stopped because it exceeded the maximum allowed wait time.

Stopped - The training job has stopped.

Stopping

- Stopping - Stopping the training job.

Valid values for SecondaryStatus are subject to change.

We no longer support the following secondary statuses:

LaunchingMLInstances

PreparingTraining

DownloadingTrainingImage

algorithmSpecification :: AlgorithmSpecification

Information about the algorithm used for training, and algorithm metadata.

resourceConfig :: ResourceConfig

Resources, including ML compute instances and ML storage volumes, that are configured for model training.

stoppingCondition :: StoppingCondition

Specifies a limit to how long a model training job can run. It also specifies how long a managed Spot training job has to complete. When the job reaches the time limit, Amazon SageMaker ends the training job. Use this API to cap model training costs.

To stop a job, Amazon SageMaker sends the algorithm the SIGTERM signal, which delays job termination for 120 seconds. Algorithms can use this 120-second window to save the model artifacts, so the results of training are not lost.

creationTime :: POSIX

A timestamp that indicates when the training job was created.

newDescribeTrainingJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeTrainingJobResponse'

-> Text
$sel:trainingJobName:DescribeTrainingJob'

-> Text
$sel:trainingJobArn:DescribeTrainingJobResponse'

-> ModelArtifacts
$sel:modelArtifacts:DescribeTrainingJobResponse'

-> TrainingJobStatus
$sel:trainingJobStatus:DescribeTrainingJobResponse'

-> SecondaryStatus
$sel:secondaryStatus:DescribeTrainingJobResponse'

-> AlgorithmSpecification
$sel:algorithmSpecification:DescribeTrainingJobResponse'

-> ResourceConfig
$sel:resourceConfig:DescribeTrainingJobResponse'

-> StoppingCondition
$sel:stoppingCondition:DescribeTrainingJobResponse'

-> UTCTime
$sel:creationTime:DescribeTrainingJobResponse'

-> DescribeTrainingJobResponse

Create a value of DescribeTrainingJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcConfig:DescribeTrainingJobResponse', describeTrainingJobResponse_vpcConfig - A VpcConfig object that specifies the VPC that this training job has access to. For more information, see Protect Training Jobs by Using an Amazon Virtual Private Cloud.

$sel:enableManagedSpotTraining:DescribeTrainingJobResponse', describeTrainingJobResponse_enableManagedSpotTraining - A Boolean indicating whether managed spot training is enabled (True) or not (False).

$sel:labelingJobArn:DescribeTrainingJobResponse', describeTrainingJobResponse_labelingJobArn - The Amazon Resource Name (ARN) of the Amazon SageMaker Ground Truth labeling job that created the transform or training job.

$sel:hyperParameters:DescribeTrainingJobResponse', describeTrainingJobResponse_hyperParameters - Algorithm-specific parameters.

$sel:debugRuleConfigurations:DescribeTrainingJobResponse', describeTrainingJobResponse_debugRuleConfigurations - Configuration information for Debugger rules for debugging output tensors.

$sel:inputDataConfig:DescribeTrainingJobResponse', describeTrainingJobResponse_inputDataConfig - An array of Channel objects that describes each data input channel.

$sel:trainingTimeInSeconds:DescribeTrainingJobResponse', describeTrainingJobResponse_trainingTimeInSeconds - The training time in seconds.

$sel:roleArn:DescribeTrainingJobResponse', describeTrainingJobResponse_roleArn - The Amazon Web Services Identity and Access Management (IAM) role configured for the training job.

$sel:experimentConfig:DescribeTrainingJobResponse', describeTrainingJobResponse_experimentConfig - Undocumented member.

$sel:enableNetworkIsolation:DescribeTrainingJobResponse', describeTrainingJobResponse_enableNetworkIsolation - If you want to allow inbound or outbound network calls, except for calls between peers within a training cluster for distributed training, choose True. If you enable network isolation for training jobs that are configured to use a VPC, Amazon SageMaker downloads and uploads customer data and model artifacts through the specified VPC, but the training container does not have network access.

$sel:profilerRuleEvaluationStatuses:DescribeTrainingJobResponse', describeTrainingJobResponse_profilerRuleEvaluationStatuses - Evaluation status of Debugger rules for profiling on a training job.

$sel:profilerConfig:DescribeTrainingJobResponse', describeTrainingJobResponse_profilerConfig - Undocumented member.

$sel:enableInterContainerTrafficEncryption:DescribeTrainingJobResponse', describeTrainingJobResponse_enableInterContainerTrafficEncryption - To encrypt all communications between ML compute instances in distributed training, choose True. Encryption provides greater security for distributed training, but training might take longer. How long it takes depends on the amount of communication between compute instances, especially if you use a deep learning algorithms in distributed training.

$sel:checkpointConfig:DescribeTrainingJobResponse', describeTrainingJobResponse_checkpointConfig - Undocumented member.

$sel:tuningJobArn:DescribeTrainingJobResponse', describeTrainingJobResponse_tuningJobArn - The Amazon Resource Name (ARN) of the associated hyperparameter tuning job if the training job was launched by a hyperparameter tuning job.

$sel:outputDataConfig:DescribeTrainingJobResponse', describeTrainingJobResponse_outputDataConfig - The S3 path where model artifacts that you configured when creating the job are stored. Amazon SageMaker creates subfolders for model artifacts.

$sel:environment:DescribeTrainingJobResponse', describeTrainingJobResponse_environment - The environment variables to set in the Docker container.

$sel:profilingStatus:DescribeTrainingJobResponse', describeTrainingJobResponse_profilingStatus - Profiling status of a training job.

$sel:finalMetricDataList:DescribeTrainingJobResponse', describeTrainingJobResponse_finalMetricDataList - A collection of MetricData objects that specify the names, values, and dates and times that the training algorithm emitted to Amazon CloudWatch.

$sel:secondaryStatusTransitions:DescribeTrainingJobResponse', describeTrainingJobResponse_secondaryStatusTransitions - A history of all of the secondary statuses that the training job has transitioned through.

$sel:autoMLJobArn:DescribeTrainingJobResponse', describeTrainingJobResponse_autoMLJobArn - The Amazon Resource Name (ARN) of an AutoML job.

$sel:profilerRuleConfigurations:DescribeTrainingJobResponse', describeTrainingJobResponse_profilerRuleConfigurations - Configuration information for Debugger rules for profiling system and framework metrics.

$sel:failureReason:DescribeTrainingJobResponse', describeTrainingJobResponse_failureReason - If the training job failed, the reason it failed.

$sel:lastModifiedTime:DescribeTrainingJobResponse', describeTrainingJobResponse_lastModifiedTime - A timestamp that indicates when the status of the training job was last modified.

$sel:tensorBoardOutputConfig:DescribeTrainingJobResponse', describeTrainingJobResponse_tensorBoardOutputConfig - Undocumented member.

$sel:debugRuleEvaluationStatuses:DescribeTrainingJobResponse', describeTrainingJobResponse_debugRuleEvaluationStatuses - Evaluation status of Debugger rules for debugging on a training job.

$sel:retryStrategy:DescribeTrainingJobResponse', describeTrainingJobResponse_retryStrategy - The number of times to retry the job when the job fails due to an InternalServerError.

$sel:debugHookConfig:DescribeTrainingJobResponse', describeTrainingJobResponse_debugHookConfig - Undocumented member.

$sel:billableTimeInSeconds:DescribeTrainingJobResponse', describeTrainingJobResponse_billableTimeInSeconds - The billable time in seconds. Billable time refers to the absolute wall-clock time.

Multiply BillableTimeInSeconds by the number of instances (InstanceCount) in your training cluster to get the total compute time Amazon SageMaker will bill you if you run distributed training. The formula is as follows: BillableTimeInSeconds * InstanceCount .

You can calculate the savings from using managed spot training using the formula (1 - BillableTimeInSeconds / TrainingTimeInSeconds) * 100. For example, if BillableTimeInSeconds is 100 and TrainingTimeInSeconds is 500, the savings is 80%.

$sel:trainingStartTime:DescribeTrainingJobResponse', describeTrainingJobResponse_trainingStartTime - Indicates the time when the training job starts on training instances. You are billed for the time interval between this time and the value of TrainingEndTime. The start time in CloudWatch Logs might be later than this time. The difference is due to the time it takes to download the training data and to the size of the training container.

$sel:trainingEndTime:DescribeTrainingJobResponse', describeTrainingJobResponse_trainingEndTime - Indicates the time when the training job ends on training instances. You are billed for the time interval between the value of TrainingStartTime and this time. For successful jobs and stopped jobs, this is the time after model artifacts are uploaded. For failed jobs, this is the time when Amazon SageMaker detects a job failure.

$sel:httpStatus:DescribeTrainingJobResponse', describeTrainingJobResponse_httpStatus - The response's http status code.

$sel:trainingJobName:DescribeTrainingJob', describeTrainingJobResponse_trainingJobName - Name of the model training job.

$sel:trainingJobArn:DescribeTrainingJobResponse', describeTrainingJobResponse_trainingJobArn - The Amazon Resource Name (ARN) of the training job.

$sel:modelArtifacts:DescribeTrainingJobResponse', describeTrainingJobResponse_modelArtifacts - Information about the Amazon S3 location that is configured for storing model artifacts.

$sel:trainingJobStatus:DescribeTrainingJobResponse', describeTrainingJobResponse_trainingJobStatus - The status of the training job.

Amazon SageMaker provides the following training job statuses:

InProgress - The training is in progress.

Completed - The training job has completed.

Failed - The training job has failed. To see the reason for the failure, see the FailureReason field in the response to a DescribeTrainingJobResponse call.

Stopping - The training job is stopping.

Stopped - The training job has stopped.

For more detailed information, see SecondaryStatus.

$sel:secondaryStatus:DescribeTrainingJobResponse', describeTrainingJobResponse_secondaryStatus - Provides detailed information about the state of the training job. For detailed information on the secondary status of the training job, see StatusMessage under SecondaryStatusTransition.

Amazon SageMaker provides primary statuses and secondary statuses that apply to each of them:

InProgress

- Starting - Starting the training job.

Downloading - An optional stage for algorithms that support File training input mode. It indicates that data is being downloaded to the ML storage volumes.

Training - Training is in progress.

Interrupted - The job stopped because the managed spot training instances were interrupted.

Uploading - Training is complete and the model artifacts are being uploaded to the S3 location.

Completed

- Completed - The training job has completed.

Failed

- Failed - The training job has failed. The reason for the failure is returned in the FailureReason field of DescribeTrainingJobResponse.

Stopped

- MaxRuntimeExceeded - The job stopped because it exceeded the maximum allowed runtime.

MaxWaitTimeExceeded - The job stopped because it exceeded the maximum allowed wait time.

Stopped - The training job has stopped.

Stopping

- Stopping - Stopping the training job.

Valid values for SecondaryStatus are subject to change.

We no longer support the following secondary statuses:

LaunchingMLInstances

PreparingTraining

DownloadingTrainingImage

$sel:algorithmSpecification:DescribeTrainingJobResponse', describeTrainingJobResponse_algorithmSpecification - Information about the algorithm used for training, and algorithm metadata.

$sel:resourceConfig:DescribeTrainingJobResponse', describeTrainingJobResponse_resourceConfig - Resources, including ML compute instances and ML storage volumes, that are configured for model training.

$sel:stoppingCondition:DescribeTrainingJobResponse', describeTrainingJobResponse_stoppingCondition - Specifies a limit to how long a model training job can run. It also specifies how long a managed Spot training job has to complete. When the job reaches the time limit, Amazon SageMaker ends the training job. Use this API to cap model training costs.

To stop a job, Amazon SageMaker sends the algorithm the SIGTERM signal, which delays job termination for 120 seconds. Algorithms can use this 120-second window to save the model artifacts, so the results of training are not lost.

$sel:creationTime:DescribeTrainingJobResponse', describeTrainingJobResponse_creationTime - A timestamp that indicates when the training job was created.

Response Lenses

describeTrainingJobResponse_vpcConfig :: Lens' DescribeTrainingJobResponse (Maybe VpcConfig)

A VpcConfig object that specifies the VPC that this training job has access to. For more information, see Protect Training Jobs by Using an Amazon Virtual Private Cloud.

describeTrainingJobResponse_enableManagedSpotTraining :: Lens' DescribeTrainingJobResponse (Maybe Bool)

A Boolean indicating whether managed spot training is enabled (True) or not (False).

describeTrainingJobResponse_labelingJobArn :: Lens' DescribeTrainingJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon SageMaker Ground Truth labeling job that created the transform or training job.

describeTrainingJobResponse_hyperParameters :: Lens' DescribeTrainingJobResponse (Maybe (HashMap Text Text))

Algorithm-specific parameters.

describeTrainingJobResponse_debugRuleConfigurations :: Lens' DescribeTrainingJobResponse (Maybe [DebugRuleConfiguration])

Configuration information for Debugger rules for debugging output tensors.

describeTrainingJobResponse_inputDataConfig :: Lens' DescribeTrainingJobResponse (Maybe (NonEmpty Channel))

An array of Channel objects that describes each data input channel.

describeTrainingJobResponse_trainingTimeInSeconds :: Lens' DescribeTrainingJobResponse (Maybe Natural)

The training time in seconds.

describeTrainingJobResponse_roleArn :: Lens' DescribeTrainingJobResponse (Maybe Text)

The Amazon Web Services Identity and Access Management (IAM) role configured for the training job.

describeTrainingJobResponse_experimentConfig :: Lens' DescribeTrainingJobResponse (Maybe ExperimentConfig)

Undocumented member.

describeTrainingJobResponse_enableNetworkIsolation :: Lens' DescribeTrainingJobResponse (Maybe Bool)

If you want to allow inbound or outbound network calls, except for calls between peers within a training cluster for distributed training, choose True. If you enable network isolation for training jobs that are configured to use a VPC, Amazon SageMaker downloads and uploads customer data and model artifacts through the specified VPC, but the training container does not have network access.

describeTrainingJobResponse_profilerRuleEvaluationStatuses :: Lens' DescribeTrainingJobResponse (Maybe [ProfilerRuleEvaluationStatus])

Evaluation status of Debugger rules for profiling on a training job.

describeTrainingJobResponse_profilerConfig :: Lens' DescribeTrainingJobResponse (Maybe ProfilerConfig)

Undocumented member.

describeTrainingJobResponse_enableInterContainerTrafficEncryption :: Lens' DescribeTrainingJobResponse (Maybe Bool)

To encrypt all communications between ML compute instances in distributed training, choose True. Encryption provides greater security for distributed training, but training might take longer. How long it takes depends on the amount of communication between compute instances, especially if you use a deep learning algorithms in distributed training.

describeTrainingJobResponse_checkpointConfig :: Lens' DescribeTrainingJobResponse (Maybe CheckpointConfig)

Undocumented member.

describeTrainingJobResponse_tuningJobArn :: Lens' DescribeTrainingJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the associated hyperparameter tuning job if the training job was launched by a hyperparameter tuning job.

describeTrainingJobResponse_outputDataConfig :: Lens' DescribeTrainingJobResponse (Maybe OutputDataConfig)

The S3 path where model artifacts that you configured when creating the job are stored. Amazon SageMaker creates subfolders for model artifacts.

describeTrainingJobResponse_environment :: Lens' DescribeTrainingJobResponse (Maybe (HashMap Text Text))

The environment variables to set in the Docker container.

describeTrainingJobResponse_profilingStatus :: Lens' DescribeTrainingJobResponse (Maybe ProfilingStatus)

Profiling status of a training job.

describeTrainingJobResponse_finalMetricDataList :: Lens' DescribeTrainingJobResponse (Maybe [MetricData])

A collection of MetricData objects that specify the names, values, and dates and times that the training algorithm emitted to Amazon CloudWatch.

describeTrainingJobResponse_secondaryStatusTransitions :: Lens' DescribeTrainingJobResponse (Maybe [SecondaryStatusTransition])

A history of all of the secondary statuses that the training job has transitioned through.

describeTrainingJobResponse_autoMLJobArn :: Lens' DescribeTrainingJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of an AutoML job.

describeTrainingJobResponse_profilerRuleConfigurations :: Lens' DescribeTrainingJobResponse (Maybe [ProfilerRuleConfiguration])

Configuration information for Debugger rules for profiling system and framework metrics.

describeTrainingJobResponse_failureReason :: Lens' DescribeTrainingJobResponse (Maybe Text)

If the training job failed, the reason it failed.

describeTrainingJobResponse_lastModifiedTime :: Lens' DescribeTrainingJobResponse (Maybe UTCTime)

A timestamp that indicates when the status of the training job was last modified.

describeTrainingJobResponse_tensorBoardOutputConfig :: Lens' DescribeTrainingJobResponse (Maybe TensorBoardOutputConfig)

Undocumented member.

describeTrainingJobResponse_debugRuleEvaluationStatuses :: Lens' DescribeTrainingJobResponse (Maybe [DebugRuleEvaluationStatus])

Evaluation status of Debugger rules for debugging on a training job.

describeTrainingJobResponse_retryStrategy :: Lens' DescribeTrainingJobResponse (Maybe RetryStrategy)

The number of times to retry the job when the job fails due to an InternalServerError.

describeTrainingJobResponse_debugHookConfig :: Lens' DescribeTrainingJobResponse (Maybe DebugHookConfig)

Undocumented member.

describeTrainingJobResponse_billableTimeInSeconds :: Lens' DescribeTrainingJobResponse (Maybe Natural)

The billable time in seconds. Billable time refers to the absolute wall-clock time.

Multiply BillableTimeInSeconds by the number of instances (InstanceCount) in your training cluster to get the total compute time Amazon SageMaker will bill you if you run distributed training. The formula is as follows: BillableTimeInSeconds * InstanceCount .

You can calculate the savings from using managed spot training using the formula (1 - BillableTimeInSeconds / TrainingTimeInSeconds) * 100. For example, if BillableTimeInSeconds is 100 and TrainingTimeInSeconds is 500, the savings is 80%.

describeTrainingJobResponse_trainingStartTime :: Lens' DescribeTrainingJobResponse (Maybe UTCTime)

Indicates the time when the training job starts on training instances. You are billed for the time interval between this time and the value of TrainingEndTime. The start time in CloudWatch Logs might be later than this time. The difference is due to the time it takes to download the training data and to the size of the training container.

describeTrainingJobResponse_trainingEndTime :: Lens' DescribeTrainingJobResponse (Maybe UTCTime)

Indicates the time when the training job ends on training instances. You are billed for the time interval between the value of TrainingStartTime and this time. For successful jobs and stopped jobs, this is the time after model artifacts are uploaded. For failed jobs, this is the time when Amazon SageMaker detects a job failure.

describeTrainingJobResponse_httpStatus :: Lens' DescribeTrainingJobResponse Int

The response's http status code.

describeTrainingJobResponse_trainingJobName :: Lens' DescribeTrainingJobResponse Text

Name of the model training job.

describeTrainingJobResponse_trainingJobArn :: Lens' DescribeTrainingJobResponse Text

The Amazon Resource Name (ARN) of the training job.

describeTrainingJobResponse_modelArtifacts :: Lens' DescribeTrainingJobResponse ModelArtifacts

Information about the Amazon S3 location that is configured for storing model artifacts.

describeTrainingJobResponse_trainingJobStatus :: Lens' DescribeTrainingJobResponse TrainingJobStatus

The status of the training job.

Amazon SageMaker provides the following training job statuses:

InProgress - The training is in progress.

Completed - The training job has completed.

Failed - The training job has failed. To see the reason for the failure, see the FailureReason field in the response to a DescribeTrainingJobResponse call.

Stopping - The training job is stopping.

Stopped - The training job has stopped.

For more detailed information, see SecondaryStatus.

describeTrainingJobResponse_secondaryStatus :: Lens' DescribeTrainingJobResponse SecondaryStatus

Provides detailed information about the state of the training job. For detailed information on the secondary status of the training job, see StatusMessage under SecondaryStatusTransition.

Amazon SageMaker provides primary statuses and secondary statuses that apply to each of them:

InProgress

- Starting - Starting the training job.

Downloading - An optional stage for algorithms that support File training input mode. It indicates that data is being downloaded to the ML storage volumes.

Training - Training is in progress.

Interrupted - The job stopped because the managed spot training instances were interrupted.

Uploading - Training is complete and the model artifacts are being uploaded to the S3 location.

Completed

- Completed - The training job has completed.

Failed

- Failed - The training job has failed. The reason for the failure is returned in the FailureReason field of DescribeTrainingJobResponse.

Stopped

- MaxRuntimeExceeded - The job stopped because it exceeded the maximum allowed runtime.

MaxWaitTimeExceeded - The job stopped because it exceeded the maximum allowed wait time.

Stopped - The training job has stopped.

Stopping

- Stopping - Stopping the training job.

Valid values for SecondaryStatus are subject to change.

We no longer support the following secondary statuses:

LaunchingMLInstances

PreparingTraining

DownloadingTrainingImage

describeTrainingJobResponse_algorithmSpecification :: Lens' DescribeTrainingJobResponse AlgorithmSpecification

Information about the algorithm used for training, and algorithm metadata.

describeTrainingJobResponse_resourceConfig :: Lens' DescribeTrainingJobResponse ResourceConfig

Resources, including ML compute instances and ML storage volumes, that are configured for model training.

describeTrainingJobResponse_stoppingCondition :: Lens' DescribeTrainingJobResponse StoppingCondition

Specifies a limit to how long a model training job can run. It also specifies how long a managed Spot training job has to complete. When the job reaches the time limit, Amazon SageMaker ends the training job. Use this API to cap model training costs.

To stop a job, Amazon SageMaker sends the algorithm the SIGTERM signal, which delays job termination for 120 seconds. Algorithms can use this 120-second window to save the model artifacts, so the results of training are not lost.

describeTrainingJobResponse_creationTime :: Lens' DescribeTrainingJobResponse UTCTime

A timestamp that indicates when the training job was created.