Network.AWS.SageMaker.DescribeTransformJob

Description

Returns information about a transform job.

Creating a Request

data DescribeTransformJob

See: newDescribeTransformJob smart constructor.

Constructors

DescribeTransformJob'

Fields

transformJobName :: Text

The name of the transform job that you want to view details of.

newDescribeTransformJob

Arguments

:: Text
$sel:transformJobName:DescribeTransformJob'

-> DescribeTransformJob

Create a value of DescribeTransformJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transformJobName:DescribeTransformJob', describeTransformJob_transformJobName - The name of the transform job that you want to view details of.

Request Lenses

describeTransformJob_transformJobName :: Lens' DescribeTransformJob Text

The name of the transform job that you want to view details of.

Destructuring the Response

data DescribeTransformJobResponse

See: newDescribeTransformJobResponse smart constructor.

Constructors

DescribeTransformJobResponse'

Fields

labelingJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon SageMaker Ground Truth labeling job that created the transform or training job.

transformStartTime :: Maybe POSIX

Indicates when the transform job starts on ML instances. You are billed for the time interval between this time and the value of TransformEndTime.

experimentConfig :: Maybe ExperimentConfig

transformOutput :: Maybe TransformOutput

Identifies the Amazon S3 location where you want Amazon SageMaker to save the results from the transform job.

maxConcurrentTransforms :: Maybe Natural

The maximum number of parallel requests on each instance node that can be launched in a transform job. The default value is 1.

environment :: Maybe (HashMap Text Text)

The environment variables to set in the Docker container. We support up to 16 key and values entries in the map.

maxPayloadInMB :: Maybe Natural

The maximum payload size, in MB, used in the transform job.

batchStrategy :: Maybe BatchStrategy

Specifies the number of records to include in a mini-batch for an HTTP inference request. A record // is a single unit of input data that inference can be made on. For example, a single line in a CSV file is a record.

To enable the batch strategy, you must set SplitType to Line, RecordIO, or TFRecord.

autoMLJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the AutoML transform job.

modelClientConfig :: Maybe ModelClientConfig

The timeout and maximum number of retries for processing a transform job invocation.

failureReason :: Maybe Text

If the transform job failed, FailureReason describes why it failed. A transform job creates a log file, which includes error messages, and stores it as an Amazon S3 object. For more information, see Log Amazon SageMaker Events with Amazon CloudWatch.

transformEndTime :: Maybe POSIX

Indicates when the transform job has been completed, or has stopped or failed. You are billed for the time interval between this time and the value of TransformStartTime.

dataProcessing :: Maybe DataProcessing

httpStatus :: Int

The response's http status code.

transformJobName :: Text

The name of the transform job.

transformJobArn :: Text

The Amazon Resource Name (ARN) of the transform job.

transformJobStatus :: TransformJobStatus

The status of the transform job. If the transform job failed, the reason is returned in the FailureReason field.

modelName :: Text

The name of the model used in the transform job.

transformInput :: TransformInput

Describes the dataset to be transformed and the Amazon S3 location where it is stored.

transformResources :: TransformResources

Describes the resources, including ML instance types and ML instance count, to use for the transform job.

creationTime :: POSIX

A timestamp that shows when the transform Job was created.

newDescribeTransformJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeTransformJobResponse'

-> Text
$sel:transformJobName:DescribeTransformJob'

-> Text
$sel:transformJobArn:DescribeTransformJobResponse'

-> TransformJobStatus
$sel:transformJobStatus:DescribeTransformJobResponse'

-> Text
$sel:modelName:DescribeTransformJobResponse'

-> TransformInput
$sel:transformInput:DescribeTransformJobResponse'

-> TransformResources
$sel:transformResources:DescribeTransformJobResponse'

-> UTCTime
$sel:creationTime:DescribeTransformJobResponse'

-> DescribeTransformJobResponse

Create a value of DescribeTransformJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:labelingJobArn:DescribeTransformJobResponse', describeTransformJobResponse_labelingJobArn - The Amazon Resource Name (ARN) of the Amazon SageMaker Ground Truth labeling job that created the transform or training job.

$sel:transformStartTime:DescribeTransformJobResponse', describeTransformJobResponse_transformStartTime - Indicates when the transform job starts on ML instances. You are billed for the time interval between this time and the value of TransformEndTime.

$sel:experimentConfig:DescribeTransformJobResponse', describeTransformJobResponse_experimentConfig - Undocumented member.

$sel:transformOutput:DescribeTransformJobResponse', describeTransformJobResponse_transformOutput - Identifies the Amazon S3 location where you want Amazon SageMaker to save the results from the transform job.

$sel:maxConcurrentTransforms:DescribeTransformJobResponse', describeTransformJobResponse_maxConcurrentTransforms - The maximum number of parallel requests on each instance node that can be launched in a transform job. The default value is 1.

$sel:environment:DescribeTransformJobResponse', describeTransformJobResponse_environment - The environment variables to set in the Docker container. We support up to 16 key and values entries in the map.

$sel:maxPayloadInMB:DescribeTransformJobResponse', describeTransformJobResponse_maxPayloadInMB - The maximum payload size, in MB, used in the transform job.

$sel:batchStrategy:DescribeTransformJobResponse', describeTransformJobResponse_batchStrategy - Specifies the number of records to include in a mini-batch for an HTTP inference request. A record // is a single unit of input data that inference can be made on. For example, a single line in a CSV file is a record.

To enable the batch strategy, you must set SplitType to Line, RecordIO, or TFRecord.

$sel:autoMLJobArn:DescribeTransformJobResponse', describeTransformJobResponse_autoMLJobArn - The Amazon Resource Name (ARN) of the AutoML transform job.

$sel:modelClientConfig:DescribeTransformJobResponse', describeTransformJobResponse_modelClientConfig - The timeout and maximum number of retries for processing a transform job invocation.

$sel:failureReason:DescribeTransformJobResponse', describeTransformJobResponse_failureReason - If the transform job failed, FailureReason describes why it failed. A transform job creates a log file, which includes error messages, and stores it as an Amazon S3 object. For more information, see Log Amazon SageMaker Events with Amazon CloudWatch.

$sel:transformEndTime:DescribeTransformJobResponse', describeTransformJobResponse_transformEndTime - Indicates when the transform job has been completed, or has stopped or failed. You are billed for the time interval between this time and the value of TransformStartTime.

$sel:dataProcessing:DescribeTransformJobResponse', describeTransformJobResponse_dataProcessing - Undocumented member.

$sel:httpStatus:DescribeTransformJobResponse', describeTransformJobResponse_httpStatus - The response's http status code.

$sel:transformJobName:DescribeTransformJob', describeTransformJobResponse_transformJobName - The name of the transform job.

$sel:transformJobArn:DescribeTransformJobResponse', describeTransformJobResponse_transformJobArn - The Amazon Resource Name (ARN) of the transform job.

$sel:transformJobStatus:DescribeTransformJobResponse', describeTransformJobResponse_transformJobStatus - The status of the transform job. If the transform job failed, the reason is returned in the FailureReason field.

$sel:modelName:DescribeTransformJobResponse', describeTransformJobResponse_modelName - The name of the model used in the transform job.

$sel:transformInput:DescribeTransformJobResponse', describeTransformJobResponse_transformInput - Describes the dataset to be transformed and the Amazon S3 location where it is stored.

$sel:transformResources:DescribeTransformJobResponse', describeTransformJobResponse_transformResources - Describes the resources, including ML instance types and ML instance count, to use for the transform job.

$sel:creationTime:DescribeTransformJobResponse', describeTransformJobResponse_creationTime - A timestamp that shows when the transform Job was created.

Response Lenses

describeTransformJobResponse_labelingJobArn :: Lens' DescribeTransformJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon SageMaker Ground Truth labeling job that created the transform or training job.

describeTransformJobResponse_transformStartTime :: Lens' DescribeTransformJobResponse (Maybe UTCTime)

Indicates when the transform job starts on ML instances. You are billed for the time interval between this time and the value of TransformEndTime.

describeTransformJobResponse_experimentConfig :: Lens' DescribeTransformJobResponse (Maybe ExperimentConfig)

Undocumented member.

describeTransformJobResponse_transformOutput :: Lens' DescribeTransformJobResponse (Maybe TransformOutput)

Identifies the Amazon S3 location where you want Amazon SageMaker to save the results from the transform job.

describeTransformJobResponse_maxConcurrentTransforms :: Lens' DescribeTransformJobResponse (Maybe Natural)

The maximum number of parallel requests on each instance node that can be launched in a transform job. The default value is 1.

describeTransformJobResponse_environment :: Lens' DescribeTransformJobResponse (Maybe (HashMap Text Text))

The environment variables to set in the Docker container. We support up to 16 key and values entries in the map.

describeTransformJobResponse_maxPayloadInMB :: Lens' DescribeTransformJobResponse (Maybe Natural)

The maximum payload size, in MB, used in the transform job.

describeTransformJobResponse_batchStrategy :: Lens' DescribeTransformJobResponse (Maybe BatchStrategy)

Specifies the number of records to include in a mini-batch for an HTTP inference request. A record // is a single unit of input data that inference can be made on. For example, a single line in a CSV file is a record.

To enable the batch strategy, you must set SplitType to Line, RecordIO, or TFRecord.

describeTransformJobResponse_autoMLJobArn :: Lens' DescribeTransformJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the AutoML transform job.

describeTransformJobResponse_modelClientConfig :: Lens' DescribeTransformJobResponse (Maybe ModelClientConfig)

The timeout and maximum number of retries for processing a transform job invocation.

describeTransformJobResponse_failureReason :: Lens' DescribeTransformJobResponse (Maybe Text)

If the transform job failed, FailureReason describes why it failed. A transform job creates a log file, which includes error messages, and stores it as an Amazon S3 object. For more information, see Log Amazon SageMaker Events with Amazon CloudWatch.

describeTransformJobResponse_transformEndTime :: Lens' DescribeTransformJobResponse (Maybe UTCTime)

Indicates when the transform job has been completed, or has stopped or failed. You are billed for the time interval between this time and the value of TransformStartTime.

describeTransformJobResponse_dataProcessing :: Lens' DescribeTransformJobResponse (Maybe DataProcessing)

Undocumented member.

describeTransformJobResponse_httpStatus :: Lens' DescribeTransformJobResponse Int

The response's http status code.

describeTransformJobResponse_transformJobName :: Lens' DescribeTransformJobResponse Text

The name of the transform job.

describeTransformJobResponse_transformJobArn :: Lens' DescribeTransformJobResponse Text

The Amazon Resource Name (ARN) of the transform job.

describeTransformJobResponse_transformJobStatus :: Lens' DescribeTransformJobResponse TransformJobStatus

The status of the transform job. If the transform job failed, the reason is returned in the FailureReason field.

describeTransformJobResponse_modelName :: Lens' DescribeTransformJobResponse Text

The name of the model used in the transform job.

describeTransformJobResponse_transformInput :: Lens' DescribeTransformJobResponse TransformInput

Describes the dataset to be transformed and the Amazon S3 location where it is stored.

describeTransformJobResponse_transformResources :: Lens' DescribeTransformJobResponse TransformResources

Describes the resources, including ML instance types and ML instance count, to use for the transform job.

describeTransformJobResponse_creationTime :: Lens' DescribeTransformJobResponse UTCTime

A timestamp that shows when the transform Job was created.