Network.AWS.SageMaker.DescribeTrialComponent

Description

Provides a list of a trials component's properties.

Creating a Request

data DescribeTrialComponent

See: newDescribeTrialComponent smart constructor.

Constructors

DescribeTrialComponent'

Fields

trialComponentName :: Text

The name of the trial component to describe.

newDescribeTrialComponent

Arguments

:: Text
$sel:trialComponentName:DescribeTrialComponent'

-> DescribeTrialComponent

Create a value of DescribeTrialComponent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trialComponentName:DescribeTrialComponent', describeTrialComponent_trialComponentName - The name of the trial component to describe.

Request Lenses

describeTrialComponent_trialComponentName :: Lens' DescribeTrialComponent Text

The name of the trial component to describe.

Destructuring the Response

data DescribeTrialComponentResponse

See: newDescribeTrialComponentResponse smart constructor.

Constructors

DescribeTrialComponentResponse'

Fields

creationTime :: Maybe POSIX

When the component was created.

status :: Maybe TrialComponentStatus

The status of the component. States include:

InProgress

Completed

Failed

metadataProperties :: Maybe MetadataProperties

trialComponentArn :: Maybe Text

The Amazon Resource Name (ARN) of the trial component.

startTime :: Maybe POSIX

When the component started.

source :: Maybe TrialComponentSource

The Amazon Resource Name (ARN) of the source and, optionally, the job type.

endTime :: Maybe POSIX

When the component ended.

metrics :: Maybe [TrialComponentMetricSummary]

The metrics for the component.

lastModifiedTime :: Maybe POSIX

When the component was last modified.

createdBy :: Maybe UserContext

Who created the trial component.

inputArtifacts :: Maybe (HashMap Text TrialComponentArtifact)

The input artifacts of the component.

lastModifiedBy :: Maybe UserContext

Who last modified the component.

displayName :: Maybe Text

The name of the component as displayed. If DisplayName isn't specified, TrialComponentName is displayed.

trialComponentName :: Maybe Text

The name of the trial component.

outputArtifacts :: Maybe (HashMap Text TrialComponentArtifact)

The output artifacts of the component.

parameters :: Maybe (HashMap Text TrialComponentParameterValue)

The hyperparameters of the component.

httpStatus :: Int

The response's http status code.

newDescribeTrialComponentResponse

Arguments

:: Int
$sel:httpStatus:DescribeTrialComponentResponse'

-> DescribeTrialComponentResponse

Create a value of DescribeTrialComponentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeTrialComponentResponse', describeTrialComponentResponse_creationTime - When the component was created.

$sel:status:DescribeTrialComponentResponse', describeTrialComponentResponse_status - The status of the component. States include:

InProgress

Completed

Failed

$sel:metadataProperties:DescribeTrialComponentResponse', describeTrialComponentResponse_metadataProperties - Undocumented member.

$sel:trialComponentArn:DescribeTrialComponentResponse', describeTrialComponentResponse_trialComponentArn - The Amazon Resource Name (ARN) of the trial component.

$sel:startTime:DescribeTrialComponentResponse', describeTrialComponentResponse_startTime - When the component started.

$sel:source:DescribeTrialComponentResponse', describeTrialComponentResponse_source - The Amazon Resource Name (ARN) of the source and, optionally, the job type.

$sel:endTime:DescribeTrialComponentResponse', describeTrialComponentResponse_endTime - When the component ended.

$sel:metrics:DescribeTrialComponentResponse', describeTrialComponentResponse_metrics - The metrics for the component.

$sel:lastModifiedTime:DescribeTrialComponentResponse', describeTrialComponentResponse_lastModifiedTime - When the component was last modified.

$sel:createdBy:DescribeTrialComponentResponse', describeTrialComponentResponse_createdBy - Who created the trial component.

$sel:inputArtifacts:DescribeTrialComponentResponse', describeTrialComponentResponse_inputArtifacts - The input artifacts of the component.

$sel:lastModifiedBy:DescribeTrialComponentResponse', describeTrialComponentResponse_lastModifiedBy - Who last modified the component.

$sel:displayName:DescribeTrialComponentResponse', describeTrialComponentResponse_displayName - The name of the component as displayed. If DisplayName isn't specified, TrialComponentName is displayed.

$sel:trialComponentName:DescribeTrialComponent', describeTrialComponentResponse_trialComponentName - The name of the trial component.

$sel:outputArtifacts:DescribeTrialComponentResponse', describeTrialComponentResponse_outputArtifacts - The output artifacts of the component.

$sel:parameters:DescribeTrialComponentResponse', describeTrialComponentResponse_parameters - The hyperparameters of the component.

$sel:httpStatus:DescribeTrialComponentResponse', describeTrialComponentResponse_httpStatus - The response's http status code.

Response Lenses

describeTrialComponentResponse_creationTime :: Lens' DescribeTrialComponentResponse (Maybe UTCTime)

When the component was created.

describeTrialComponentResponse_status :: Lens' DescribeTrialComponentResponse (Maybe TrialComponentStatus)

The status of the component. States include:

InProgress

Completed

Failed

describeTrialComponentResponse_metadataProperties :: Lens' DescribeTrialComponentResponse (Maybe MetadataProperties)

Undocumented member.

describeTrialComponentResponse_trialComponentArn :: Lens' DescribeTrialComponentResponse (Maybe Text)

The Amazon Resource Name (ARN) of the trial component.

describeTrialComponentResponse_startTime :: Lens' DescribeTrialComponentResponse (Maybe UTCTime)

When the component started.

describeTrialComponentResponse_source :: Lens' DescribeTrialComponentResponse (Maybe TrialComponentSource)

The Amazon Resource Name (ARN) of the source and, optionally, the job type.

describeTrialComponentResponse_endTime :: Lens' DescribeTrialComponentResponse (Maybe UTCTime)

When the component ended.

describeTrialComponentResponse_metrics :: Lens' DescribeTrialComponentResponse (Maybe [TrialComponentMetricSummary])

The metrics for the component.

describeTrialComponentResponse_lastModifiedTime :: Lens' DescribeTrialComponentResponse (Maybe UTCTime)

When the component was last modified.

describeTrialComponentResponse_createdBy :: Lens' DescribeTrialComponentResponse (Maybe UserContext)

Who created the trial component.

describeTrialComponentResponse_inputArtifacts :: Lens' DescribeTrialComponentResponse (Maybe (HashMap Text TrialComponentArtifact))

The input artifacts of the component.

describeTrialComponentResponse_lastModifiedBy :: Lens' DescribeTrialComponentResponse (Maybe UserContext)

Who last modified the component.

describeTrialComponentResponse_displayName :: Lens' DescribeTrialComponentResponse (Maybe Text)

The name of the component as displayed. If DisplayName isn't specified, TrialComponentName is displayed.

describeTrialComponentResponse_trialComponentName :: Lens' DescribeTrialComponentResponse (Maybe Text)

The name of the trial component.

describeTrialComponentResponse_outputArtifacts :: Lens' DescribeTrialComponentResponse (Maybe (HashMap Text TrialComponentArtifact))

The output artifacts of the component.

describeTrialComponentResponse_parameters :: Lens' DescribeTrialComponentResponse (Maybe (HashMap Text TrialComponentParameterValue))

The hyperparameters of the component.

describeTrialComponentResponse_httpStatus :: Lens' DescribeTrialComponentResponse Int

The response's http status code.