Network.AWS.SageMaker.DescribeUserProfile

Description

Describes a user profile. For more information, see CreateUserProfile.

Creating a Request

data DescribeUserProfile

See: newDescribeUserProfile smart constructor.

Constructors

DescribeUserProfile'

Fields

domainId :: Text

The domain ID.

userProfileName :: Text

The user profile name. This value is not case sensitive.

newDescribeUserProfile

Arguments

:: Text
$sel:domainId:DescribeUserProfile'

-> Text
$sel:userProfileName:DescribeUserProfile'

-> DescribeUserProfile

Create a value of DescribeUserProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainId:DescribeUserProfile', describeUserProfile_domainId - The domain ID.

$sel:userProfileName:DescribeUserProfile', describeUserProfile_userProfileName - The user profile name. This value is not case sensitive.

Request Lenses

describeUserProfile_domainId :: Lens' DescribeUserProfile Text

The domain ID.

describeUserProfile_userProfileName :: Lens' DescribeUserProfile Text

The user profile name. This value is not case sensitive.

Destructuring the Response

data DescribeUserProfileResponse

See: newDescribeUserProfileResponse smart constructor.

Constructors

DescribeUserProfileResponse'

Fields

creationTime :: Maybe POSIX

The creation time.

userSettings :: Maybe UserSettings

A collection of settings.

status :: Maybe UserProfileStatus

The status.

failureReason :: Maybe Text

The failure reason.

singleSignOnUserValue :: Maybe Text

The SSO user value.

userProfileName :: Maybe Text

The user profile name.

lastModifiedTime :: Maybe POSIX

The last modified time.

homeEfsFileSystemUid :: Maybe Text

The ID of the user's profile in the Amazon Elastic File System (EFS) volume.

userProfileArn :: Maybe Text

The user profile Amazon Resource Name (ARN).

singleSignOnUserIdentifier :: Maybe Text

The SSO user identifier.

domainId :: Maybe Text

The ID of the domain that contains the profile.

httpStatus :: Int

The response's http status code.

newDescribeUserProfileResponse

Arguments

:: Int
$sel:httpStatus:DescribeUserProfileResponse'

-> DescribeUserProfileResponse

Create a value of DescribeUserProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeUserProfileResponse', describeUserProfileResponse_creationTime - The creation time.

$sel:userSettings:DescribeUserProfileResponse', describeUserProfileResponse_userSettings - A collection of settings.

$sel:status:DescribeUserProfileResponse', describeUserProfileResponse_status - The status.

$sel:failureReason:DescribeUserProfileResponse', describeUserProfileResponse_failureReason - The failure reason.

$sel:singleSignOnUserValue:DescribeUserProfileResponse', describeUserProfileResponse_singleSignOnUserValue - The SSO user value.

$sel:userProfileName:DescribeUserProfile', describeUserProfileResponse_userProfileName - The user profile name.

$sel:lastModifiedTime:DescribeUserProfileResponse', describeUserProfileResponse_lastModifiedTime - The last modified time.

$sel:homeEfsFileSystemUid:DescribeUserProfileResponse', describeUserProfileResponse_homeEfsFileSystemUid - The ID of the user's profile in the Amazon Elastic File System (EFS) volume.

$sel:userProfileArn:DescribeUserProfileResponse', describeUserProfileResponse_userProfileArn - The user profile Amazon Resource Name (ARN).

$sel:singleSignOnUserIdentifier:DescribeUserProfileResponse', describeUserProfileResponse_singleSignOnUserIdentifier - The SSO user identifier.

$sel:domainId:DescribeUserProfile', describeUserProfileResponse_domainId - The ID of the domain that contains the profile.

$sel:httpStatus:DescribeUserProfileResponse', describeUserProfileResponse_httpStatus - The response's http status code.

Response Lenses

describeUserProfileResponse_creationTime :: Lens' DescribeUserProfileResponse (Maybe UTCTime)

The creation time.

describeUserProfileResponse_userSettings :: Lens' DescribeUserProfileResponse (Maybe UserSettings)

A collection of settings.

describeUserProfileResponse_status :: Lens' DescribeUserProfileResponse (Maybe UserProfileStatus)

The status.

describeUserProfileResponse_failureReason :: Lens' DescribeUserProfileResponse (Maybe Text)

The failure reason.

describeUserProfileResponse_singleSignOnUserValue :: Lens' DescribeUserProfileResponse (Maybe Text)

The SSO user value.

describeUserProfileResponse_userProfileName :: Lens' DescribeUserProfileResponse (Maybe Text)

The user profile name.

describeUserProfileResponse_lastModifiedTime :: Lens' DescribeUserProfileResponse (Maybe UTCTime)

The last modified time.

describeUserProfileResponse_homeEfsFileSystemUid :: Lens' DescribeUserProfileResponse (Maybe Text)

The ID of the user's profile in the Amazon Elastic File System (EFS) volume.

describeUserProfileResponse_userProfileArn :: Lens' DescribeUserProfileResponse (Maybe Text)

The user profile Amazon Resource Name (ARN).

describeUserProfileResponse_singleSignOnUserIdentifier :: Lens' DescribeUserProfileResponse (Maybe Text)

The SSO user identifier.

describeUserProfileResponse_domainId :: Lens' DescribeUserProfileResponse (Maybe Text)

The ID of the domain that contains the profile.

describeUserProfileResponse_httpStatus :: Lens' DescribeUserProfileResponse Int

The response's http status code.