Network.AWS.SageMaker.DescribeWorkforce

Description

Lists private workforce information, including workforce name, Amazon Resource Name (ARN), and, if applicable, allowed IP address ranges (CIDRs). Allowable IP address ranges are the IP addresses that workers can use to access tasks.

This operation applies only to private workforces.

Creating a Request

data DescribeWorkforce

See: newDescribeWorkforce smart constructor.

Constructors

DescribeWorkforce'

Fields

workforceName :: Text

The name of the private workforce whose access you want to restrict. WorkforceName is automatically set to default when a workforce is created and cannot be modified.

newDescribeWorkforce

Arguments

:: Text
$sel:workforceName:DescribeWorkforce'

-> DescribeWorkforce

Create a value of DescribeWorkforce with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workforceName:DescribeWorkforce', describeWorkforce_workforceName - The name of the private workforce whose access you want to restrict. WorkforceName is automatically set to default when a workforce is created and cannot be modified.

Request Lenses

describeWorkforce_workforceName :: Lens' DescribeWorkforce Text

The name of the private workforce whose access you want to restrict. WorkforceName is automatically set to default when a workforce is created and cannot be modified.

Destructuring the Response

data DescribeWorkforceResponse

See: newDescribeWorkforceResponse smart constructor.

Constructors

DescribeWorkforceResponse'

Fields

httpStatus :: Int

The response's http status code.

workforce :: Workforce

A single private workforce, which is automatically created when you create your first private work team. You can create one private work force in each Amazon Web Services Region. By default, any workforce-related API operation used in a specific region will apply to the workforce created in that region. To learn how to create a private workforce, see Create a Private Workforce.

newDescribeWorkforceResponse

Arguments

:: Int
$sel:httpStatus:DescribeWorkforceResponse'

-> Workforce
$sel:workforce:DescribeWorkforceResponse'

-> DescribeWorkforceResponse

Create a value of DescribeWorkforceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeWorkforceResponse', describeWorkforceResponse_httpStatus - The response's http status code.

$sel:workforce:DescribeWorkforceResponse', describeWorkforceResponse_workforce - A single private workforce, which is automatically created when you create your first private work team. You can create one private work force in each Amazon Web Services Region. By default, any workforce-related API operation used in a specific region will apply to the workforce created in that region. To learn how to create a private workforce, see Create a Private Workforce.

Response Lenses

describeWorkforceResponse_httpStatus :: Lens' DescribeWorkforceResponse Int

The response's http status code.

describeWorkforceResponse_workforce :: Lens' DescribeWorkforceResponse Workforce

A single private workforce, which is automatically created when you create your first private work team. You can create one private work force in each Amazon Web Services Region. By default, any workforce-related API operation used in a specific region will apply to the workforce created in that region. To learn how to create a private workforce, see Create a Private Workforce.