Network.AWS.SageMaker.ListActions

Description

Lists the actions in your account and their properties.

This operation returns paginated results.

Creating a Request

data ListActions

See: newListActions smart constructor.

Constructors

ListActions'

Fields

createdAfter :: Maybe POSIX

A filter that returns only actions created on or after the specified time.

nextToken :: Maybe Text

If the previous call to ListActions didn't return the full set of actions, the call returns a token for getting the next set of actions.

sortOrder :: Maybe SortOrder

The sort order. The default value is Descending.

sourceUri :: Maybe Text

A filter that returns only actions with the specified source URI.

actionType :: Maybe Text

A filter that returns only actions of the specified type.

maxResults :: Maybe Natural

The maximum number of actions to return in the response. The default value is 10.

createdBefore :: Maybe POSIX

A filter that returns only actions created on or before the specified time.

sortBy :: Maybe SortActionsBy

The property used to sort results. The default value is CreationTime.

newListActions :: ListActions

Create a value of ListActions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAfter:ListActions', listActions_createdAfter - A filter that returns only actions created on or after the specified time.

$sel:nextToken:ListActions', listActions_nextToken - If the previous call to ListActions didn't return the full set of actions, the call returns a token for getting the next set of actions.

$sel:sortOrder:ListActions', listActions_sortOrder - The sort order. The default value is Descending.

$sel:sourceUri:ListActions', listActions_sourceUri - A filter that returns only actions with the specified source URI.

$sel:actionType:ListActions', listActions_actionType - A filter that returns only actions of the specified type.

$sel:maxResults:ListActions', listActions_maxResults - The maximum number of actions to return in the response. The default value is 10.

$sel:createdBefore:ListActions', listActions_createdBefore - A filter that returns only actions created on or before the specified time.

$sel:sortBy:ListActions', listActions_sortBy - The property used to sort results. The default value is CreationTime.

Request Lenses

listActions_createdAfter :: Lens' ListActions (Maybe UTCTime)

A filter that returns only actions created on or after the specified time.

listActions_nextToken :: Lens' ListActions (Maybe Text)

If the previous call to ListActions didn't return the full set of actions, the call returns a token for getting the next set of actions.

listActions_sortOrder :: Lens' ListActions (Maybe SortOrder)

The sort order. The default value is Descending.

listActions_sourceUri :: Lens' ListActions (Maybe Text)

A filter that returns only actions with the specified source URI.

listActions_actionType :: Lens' ListActions (Maybe Text)

A filter that returns only actions of the specified type.

listActions_maxResults :: Lens' ListActions (Maybe Natural)

The maximum number of actions to return in the response. The default value is 10.

listActions_createdBefore :: Lens' ListActions (Maybe UTCTime)

A filter that returns only actions created on or before the specified time.

listActions_sortBy :: Lens' ListActions (Maybe SortActionsBy)

The property used to sort results. The default value is CreationTime.

Destructuring the Response

data ListActionsResponse

See: newListActionsResponse smart constructor.

Constructors

ListActionsResponse'

Fields

actionSummaries :: Maybe [ActionSummary]

A list of actions and their properties.

nextToken :: Maybe Text

A token for getting the next set of actions, if there are any.

httpStatus :: Int

The response's http status code.

newListActionsResponse

Arguments

:: Int
$sel:httpStatus:ListActionsResponse'

-> ListActionsResponse

Create a value of ListActionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actionSummaries:ListActionsResponse', listActionsResponse_actionSummaries - A list of actions and their properties.

$sel:nextToken:ListActions', listActionsResponse_nextToken - A token for getting the next set of actions, if there are any.

$sel:httpStatus:ListActionsResponse', listActionsResponse_httpStatus - The response's http status code.

Response Lenses

listActionsResponse_actionSummaries :: Lens' ListActionsResponse (Maybe [ActionSummary])

A list of actions and their properties.

listActionsResponse_nextToken :: Lens' ListActionsResponse (Maybe Text)

A token for getting the next set of actions, if there are any.

listActionsResponse_httpStatus :: Lens' ListActionsResponse Int

The response's http status code.