Network.AWS.SageMaker.ListAlgorithms

Description

Lists the machine learning algorithms that have been created.

This operation returns paginated results.

Creating a Request

data ListAlgorithms

See: newListAlgorithms smart constructor.

Constructors

ListAlgorithms'

Fields

nextToken :: Maybe Text

If the response to a previous ListAlgorithms request was truncated, the response includes a NextToken. To retrieve the next set of algorithms, use the token in the next request.

sortOrder :: Maybe SortOrder

The sort order for the results. The default is Ascending.

nameContains :: Maybe Text

A string in the algorithm name. This filter returns only algorithms whose name contains the specified string.

maxResults :: Maybe Natural

The maximum number of algorithms to return in the response.

creationTimeBefore :: Maybe POSIX

A filter that returns only algorithms created before the specified time (timestamp).

sortBy :: Maybe AlgorithmSortBy

The parameter by which to sort the results. The default is CreationTime.

creationTimeAfter :: Maybe POSIX

A filter that returns only algorithms created after the specified time (timestamp).

newListAlgorithms :: ListAlgorithms

Create a value of ListAlgorithms with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAlgorithms', listAlgorithms_nextToken - If the response to a previous ListAlgorithms request was truncated, the response includes a NextToken. To retrieve the next set of algorithms, use the token in the next request.

$sel:sortOrder:ListAlgorithms', listAlgorithms_sortOrder - The sort order for the results. The default is Ascending.

$sel:nameContains:ListAlgorithms', listAlgorithms_nameContains - A string in the algorithm name. This filter returns only algorithms whose name contains the specified string.

$sel:maxResults:ListAlgorithms', listAlgorithms_maxResults - The maximum number of algorithms to return in the response.

$sel:creationTimeBefore:ListAlgorithms', listAlgorithms_creationTimeBefore - A filter that returns only algorithms created before the specified time (timestamp).

$sel:sortBy:ListAlgorithms', listAlgorithms_sortBy - The parameter by which to sort the results. The default is CreationTime.

$sel:creationTimeAfter:ListAlgorithms', listAlgorithms_creationTimeAfter - A filter that returns only algorithms created after the specified time (timestamp).

Request Lenses

listAlgorithms_nextToken :: Lens' ListAlgorithms (Maybe Text)

If the response to a previous ListAlgorithms request was truncated, the response includes a NextToken. To retrieve the next set of algorithms, use the token in the next request.

listAlgorithms_sortOrder :: Lens' ListAlgorithms (Maybe SortOrder)

The sort order for the results. The default is Ascending.

listAlgorithms_nameContains :: Lens' ListAlgorithms (Maybe Text)

A string in the algorithm name. This filter returns only algorithms whose name contains the specified string.

listAlgorithms_maxResults :: Lens' ListAlgorithms (Maybe Natural)

The maximum number of algorithms to return in the response.

listAlgorithms_creationTimeBefore :: Lens' ListAlgorithms (Maybe UTCTime)

A filter that returns only algorithms created before the specified time (timestamp).

listAlgorithms_sortBy :: Lens' ListAlgorithms (Maybe AlgorithmSortBy)

The parameter by which to sort the results. The default is CreationTime.

listAlgorithms_creationTimeAfter :: Lens' ListAlgorithms (Maybe UTCTime)

A filter that returns only algorithms created after the specified time (timestamp).

Destructuring the Response

data ListAlgorithmsResponse

See: newListAlgorithmsResponse smart constructor.

Constructors

ListAlgorithmsResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of algorithms, use it in the subsequent request.

httpStatus :: Int

The response's http status code.

algorithmSummaryList :: [AlgorithmSummary]

An array of @AlgorithmSummary@ objects, each of which lists an

algorithm.

newListAlgorithmsResponse

Arguments

:: Int
$sel:httpStatus:ListAlgorithmsResponse'

-> ListAlgorithmsResponse

Create a value of ListAlgorithmsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAlgorithms', listAlgorithmsResponse_nextToken - If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of algorithms, use it in the subsequent request.

$sel:httpStatus:ListAlgorithmsResponse', listAlgorithmsResponse_httpStatus - The response's http status code.

$sel:algorithmSummaryList:ListAlgorithmsResponse', listAlgorithmsResponse_algorithmSummaryList - >An array of AlgorithmSummary objects, each of which lists an algorithm.

Response Lenses

listAlgorithmsResponse_nextToken :: Lens' ListAlgorithmsResponse (Maybe Text)

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of algorithms, use it in the subsequent request.

listAlgorithmsResponse_httpStatus :: Lens' ListAlgorithmsResponse Int

The response's http status code.

listAlgorithmsResponse_algorithmSummaryList :: Lens' ListAlgorithmsResponse [AlgorithmSummary]

An array of @AlgorithmSummary@ objects, each of which lists an

algorithm.