Network.AWS.SageMaker.ListAppImageConfigs

Description

Lists the AppImageConfigs in your account and their properties. The list can be filtered by creation time or modified time, and whether the AppImageConfig name contains a specified string.

This operation returns paginated results.

Creating a Request

data ListAppImageConfigs

See: newListAppImageConfigs smart constructor.

Constructors

ListAppImageConfigs'

Fields

nextToken :: Maybe Text

If the previous call to ListImages didn't return the full set of AppImageConfigs, the call returns a token for getting the next set of AppImageConfigs.

sortOrder :: Maybe SortOrder

The sort order. The default value is Descending.

nameContains :: Maybe Text

A filter that returns only AppImageConfigs whose name contains the specified string.

maxResults :: Maybe Natural

The maximum number of AppImageConfigs to return in the response. The default value is 10.

creationTimeBefore :: Maybe POSIX

A filter that returns only AppImageConfigs created on or before the specified time.

modifiedTimeBefore :: Maybe POSIX

A filter that returns only AppImageConfigs modified on or before the specified time.

sortBy :: Maybe AppImageConfigSortKey

The property used to sort results. The default value is CreationTime.

modifiedTimeAfter :: Maybe POSIX

A filter that returns only AppImageConfigs modified on or after the specified time.

creationTimeAfter :: Maybe POSIX

A filter that returns only AppImageConfigs created on or after the specified time.

newListAppImageConfigs :: ListAppImageConfigs

Create a value of ListAppImageConfigs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAppImageConfigs', listAppImageConfigs_nextToken - If the previous call to ListImages didn't return the full set of AppImageConfigs, the call returns a token for getting the next set of AppImageConfigs.

$sel:sortOrder:ListAppImageConfigs', listAppImageConfigs_sortOrder - The sort order. The default value is Descending.

$sel:nameContains:ListAppImageConfigs', listAppImageConfigs_nameContains - A filter that returns only AppImageConfigs whose name contains the specified string.

$sel:maxResults:ListAppImageConfigs', listAppImageConfigs_maxResults - The maximum number of AppImageConfigs to return in the response. The default value is 10.

$sel:creationTimeBefore:ListAppImageConfigs', listAppImageConfigs_creationTimeBefore - A filter that returns only AppImageConfigs created on or before the specified time.

$sel:modifiedTimeBefore:ListAppImageConfigs', listAppImageConfigs_modifiedTimeBefore - A filter that returns only AppImageConfigs modified on or before the specified time.

$sel:sortBy:ListAppImageConfigs', listAppImageConfigs_sortBy - The property used to sort results. The default value is CreationTime.

$sel:modifiedTimeAfter:ListAppImageConfigs', listAppImageConfigs_modifiedTimeAfter - A filter that returns only AppImageConfigs modified on or after the specified time.

$sel:creationTimeAfter:ListAppImageConfigs', listAppImageConfigs_creationTimeAfter - A filter that returns only AppImageConfigs created on or after the specified time.

Request Lenses

listAppImageConfigs_nextToken :: Lens' ListAppImageConfigs (Maybe Text)

If the previous call to ListImages didn't return the full set of AppImageConfigs, the call returns a token for getting the next set of AppImageConfigs.

listAppImageConfigs_sortOrder :: Lens' ListAppImageConfigs (Maybe SortOrder)

The sort order. The default value is Descending.

listAppImageConfigs_nameContains :: Lens' ListAppImageConfigs (Maybe Text)

A filter that returns only AppImageConfigs whose name contains the specified string.

listAppImageConfigs_maxResults :: Lens' ListAppImageConfigs (Maybe Natural)

The maximum number of AppImageConfigs to return in the response. The default value is 10.

listAppImageConfigs_creationTimeBefore :: Lens' ListAppImageConfigs (Maybe UTCTime)

A filter that returns only AppImageConfigs created on or before the specified time.

listAppImageConfigs_modifiedTimeBefore :: Lens' ListAppImageConfigs (Maybe UTCTime)

A filter that returns only AppImageConfigs modified on or before the specified time.

listAppImageConfigs_sortBy :: Lens' ListAppImageConfigs (Maybe AppImageConfigSortKey)

The property used to sort results. The default value is CreationTime.

listAppImageConfigs_modifiedTimeAfter :: Lens' ListAppImageConfigs (Maybe UTCTime)

A filter that returns only AppImageConfigs modified on or after the specified time.

listAppImageConfigs_creationTimeAfter :: Lens' ListAppImageConfigs (Maybe UTCTime)

A filter that returns only AppImageConfigs created on or after the specified time.

Destructuring the Response

data ListAppImageConfigsResponse

See: newListAppImageConfigsResponse smart constructor.

Constructors

ListAppImageConfigsResponse'

Fields

nextToken :: Maybe Text

A token for getting the next set of AppImageConfigs, if there are any.

appImageConfigs :: Maybe [AppImageConfigDetails]

A list of AppImageConfigs and their properties.

httpStatus :: Int

The response's http status code.

newListAppImageConfigsResponse

Arguments

:: Int
$sel:httpStatus:ListAppImageConfigsResponse'

-> ListAppImageConfigsResponse

Create a value of ListAppImageConfigsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAppImageConfigs', listAppImageConfigsResponse_nextToken - A token for getting the next set of AppImageConfigs, if there are any.

$sel:appImageConfigs:ListAppImageConfigsResponse', listAppImageConfigsResponse_appImageConfigs - A list of AppImageConfigs and their properties.

$sel:httpStatus:ListAppImageConfigsResponse', listAppImageConfigsResponse_httpStatus - The response's http status code.

Response Lenses

listAppImageConfigsResponse_nextToken :: Lens' ListAppImageConfigsResponse (Maybe Text)

A token for getting the next set of AppImageConfigs, if there are any.

listAppImageConfigsResponse_appImageConfigs :: Lens' ListAppImageConfigsResponse (Maybe [AppImageConfigDetails])

A list of AppImageConfigs and their properties.

listAppImageConfigsResponse_httpStatus :: Lens' ListAppImageConfigsResponse Int

The response's http status code.