Network.AWS.SageMaker.ListArtifacts

Description

Lists the artifacts in your account and their properties.

This operation returns paginated results.

Creating a Request

data ListArtifacts

See: newListArtifacts smart constructor.

Constructors

ListArtifacts'

Fields

createdAfter :: Maybe POSIX

A filter that returns only artifacts created on or after the specified time.

nextToken :: Maybe Text

If the previous call to ListArtifacts didn't return the full set of artifacts, the call returns a token for getting the next set of artifacts.

sortOrder :: Maybe SortOrder

The sort order. The default value is Descending.

sourceUri :: Maybe Text

A filter that returns only artifacts with the specified source URI.

artifactType :: Maybe Text

A filter that returns only artifacts of the specified type.

maxResults :: Maybe Natural

The maximum number of artifacts to return in the response. The default value is 10.

createdBefore :: Maybe POSIX

A filter that returns only artifacts created on or before the specified time.

sortBy :: Maybe SortArtifactsBy

The property used to sort results. The default value is CreationTime.

newListArtifacts :: ListArtifacts

Create a value of ListArtifacts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAfter:ListArtifacts', listArtifacts_createdAfter - A filter that returns only artifacts created on or after the specified time.

$sel:nextToken:ListArtifacts', listArtifacts_nextToken - If the previous call to ListArtifacts didn't return the full set of artifacts, the call returns a token for getting the next set of artifacts.

$sel:sortOrder:ListArtifacts', listArtifacts_sortOrder - The sort order. The default value is Descending.

$sel:sourceUri:ListArtifacts', listArtifacts_sourceUri - A filter that returns only artifacts with the specified source URI.

$sel:artifactType:ListArtifacts', listArtifacts_artifactType - A filter that returns only artifacts of the specified type.

$sel:maxResults:ListArtifacts', listArtifacts_maxResults - The maximum number of artifacts to return in the response. The default value is 10.

$sel:createdBefore:ListArtifacts', listArtifacts_createdBefore - A filter that returns only artifacts created on or before the specified time.

$sel:sortBy:ListArtifacts', listArtifacts_sortBy - The property used to sort results. The default value is CreationTime.

Request Lenses

listArtifacts_createdAfter :: Lens' ListArtifacts (Maybe UTCTime)

A filter that returns only artifacts created on or after the specified time.

listArtifacts_nextToken :: Lens' ListArtifacts (Maybe Text)

If the previous call to ListArtifacts didn't return the full set of artifacts, the call returns a token for getting the next set of artifacts.

listArtifacts_sortOrder :: Lens' ListArtifacts (Maybe SortOrder)

The sort order. The default value is Descending.

listArtifacts_sourceUri :: Lens' ListArtifacts (Maybe Text)

A filter that returns only artifacts with the specified source URI.

listArtifacts_artifactType :: Lens' ListArtifacts (Maybe Text)

A filter that returns only artifacts of the specified type.

listArtifacts_maxResults :: Lens' ListArtifacts (Maybe Natural)

The maximum number of artifacts to return in the response. The default value is 10.

listArtifacts_createdBefore :: Lens' ListArtifacts (Maybe UTCTime)

A filter that returns only artifacts created on or before the specified time.

listArtifacts_sortBy :: Lens' ListArtifacts (Maybe SortArtifactsBy)

The property used to sort results. The default value is CreationTime.

Destructuring the Response

data ListArtifactsResponse

See: newListArtifactsResponse smart constructor.

Constructors

ListArtifactsResponse'

Fields

artifactSummaries :: Maybe [ArtifactSummary]

A list of artifacts and their properties.

nextToken :: Maybe Text

A token for getting the next set of artifacts, if there are any.

httpStatus :: Int

The response's http status code.

newListArtifactsResponse

Arguments

:: Int
$sel:httpStatus:ListArtifactsResponse'

-> ListArtifactsResponse

Create a value of ListArtifactsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:artifactSummaries:ListArtifactsResponse', listArtifactsResponse_artifactSummaries - A list of artifacts and their properties.

$sel:nextToken:ListArtifacts', listArtifactsResponse_nextToken - A token for getting the next set of artifacts, if there are any.

$sel:httpStatus:ListArtifactsResponse', listArtifactsResponse_httpStatus - The response's http status code.

Response Lenses

listArtifactsResponse_artifactSummaries :: Lens' ListArtifactsResponse (Maybe [ArtifactSummary])

A list of artifacts and their properties.

listArtifactsResponse_nextToken :: Lens' ListArtifactsResponse (Maybe Text)

A token for getting the next set of artifacts, if there are any.

listArtifactsResponse_httpStatus :: Lens' ListArtifactsResponse Int

The response's http status code.