Network.AWS.SageMaker.ListAssociations

Description

Lists the associations in your account and their properties.

This operation returns paginated results.

Creating a Request

data ListAssociations

See: newListAssociations smart constructor.

Constructors

ListAssociations'

Fields

destinationType :: Maybe Text

A filter that returns only associations with the specified destination type.

nextToken :: Maybe Text

If the previous call to ListAssociations didn't return the full set of associations, the call returns a token for getting the next set of associations.

sortOrder :: Maybe SortOrder

The sort order. The default value is Descending.

createdAfter :: Maybe POSIX

A filter that returns only associations created on or after the specified time.

createdBefore :: Maybe POSIX

A filter that returns only associations created on or before the specified time.

destinationArn :: Maybe Text

A filter that returns only associations with the specified destination Amazon Resource Name (ARN).

maxResults :: Maybe Natural

The maximum number of associations to return in the response. The default value is 10.

associationType :: Maybe AssociationEdgeType

A filter that returns only associations of the specified type.

sortBy :: Maybe SortAssociationsBy

The property used to sort results. The default value is CreationTime.

sourceArn :: Maybe Text

A filter that returns only associations with the specified source ARN.

sourceType :: Maybe Text

A filter that returns only associations with the specified source type.

newListAssociations :: ListAssociations

Create a value of ListAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationType:ListAssociations', listAssociations_destinationType - A filter that returns only associations with the specified destination type.

$sel:nextToken:ListAssociations', listAssociations_nextToken - If the previous call to ListAssociations didn't return the full set of associations, the call returns a token for getting the next set of associations.

$sel:sortOrder:ListAssociations', listAssociations_sortOrder - The sort order. The default value is Descending.

$sel:createdAfter:ListAssociations', listAssociations_createdAfter - A filter that returns only associations created on or after the specified time.

$sel:createdBefore:ListAssociations', listAssociations_createdBefore - A filter that returns only associations created on or before the specified time.

$sel:destinationArn:ListAssociations', listAssociations_destinationArn - A filter that returns only associations with the specified destination Amazon Resource Name (ARN).

$sel:maxResults:ListAssociations', listAssociations_maxResults - The maximum number of associations to return in the response. The default value is 10.

$sel:associationType:ListAssociations', listAssociations_associationType - A filter that returns only associations of the specified type.

$sel:sortBy:ListAssociations', listAssociations_sortBy - The property used to sort results. The default value is CreationTime.

$sel:sourceArn:ListAssociations', listAssociations_sourceArn - A filter that returns only associations with the specified source ARN.

$sel:sourceType:ListAssociations', listAssociations_sourceType - A filter that returns only associations with the specified source type.

Request Lenses

listAssociations_destinationType :: Lens' ListAssociations (Maybe Text)

A filter that returns only associations with the specified destination type.

listAssociations_nextToken :: Lens' ListAssociations (Maybe Text)

If the previous call to ListAssociations didn't return the full set of associations, the call returns a token for getting the next set of associations.

listAssociations_sortOrder :: Lens' ListAssociations (Maybe SortOrder)

The sort order. The default value is Descending.

listAssociations_createdAfter :: Lens' ListAssociations (Maybe UTCTime)

A filter that returns only associations created on or after the specified time.

listAssociations_createdBefore :: Lens' ListAssociations (Maybe UTCTime)

A filter that returns only associations created on or before the specified time.

listAssociations_destinationArn :: Lens' ListAssociations (Maybe Text)

A filter that returns only associations with the specified destination Amazon Resource Name (ARN).

listAssociations_maxResults :: Lens' ListAssociations (Maybe Natural)

The maximum number of associations to return in the response. The default value is 10.

listAssociations_associationType :: Lens' ListAssociations (Maybe AssociationEdgeType)

A filter that returns only associations of the specified type.

listAssociations_sortBy :: Lens' ListAssociations (Maybe SortAssociationsBy)

The property used to sort results. The default value is CreationTime.

listAssociations_sourceArn :: Lens' ListAssociations (Maybe Text)

A filter that returns only associations with the specified source ARN.

listAssociations_sourceType :: Lens' ListAssociations (Maybe Text)

A filter that returns only associations with the specified source type.

Destructuring the Response

data ListAssociationsResponse

See: newListAssociationsResponse smart constructor.

Constructors

ListAssociationsResponse'

Fields

nextToken :: Maybe Text

A token for getting the next set of associations, if there are any.

associationSummaries :: Maybe [AssociationSummary]

A list of associations and their properties.

httpStatus :: Int

The response's http status code.

newListAssociationsResponse

Arguments

:: Int
$sel:httpStatus:ListAssociationsResponse'

-> ListAssociationsResponse

Create a value of ListAssociationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssociations', listAssociationsResponse_nextToken - A token for getting the next set of associations, if there are any.

$sel:associationSummaries:ListAssociationsResponse', listAssociationsResponse_associationSummaries - A list of associations and their properties.

$sel:httpStatus:ListAssociationsResponse', listAssociationsResponse_httpStatus - The response's http status code.

Response Lenses

listAssociationsResponse_nextToken :: Lens' ListAssociationsResponse (Maybe Text)

A token for getting the next set of associations, if there are any.

listAssociationsResponse_associationSummaries :: Lens' ListAssociationsResponse (Maybe [AssociationSummary])

A list of associations and their properties.

listAssociationsResponse_httpStatus :: Lens' ListAssociationsResponse Int

The response's http status code.