Network.AWS.SageMaker.ListAutoMLJobs

Description

Request a list of jobs.

This operation returns paginated results.

Creating a Request

data ListAutoMLJobs

See: newListAutoMLJobs smart constructor.

Constructors

ListAutoMLJobs'

Fields

lastModifiedTimeBefore :: Maybe POSIX

Request a list of jobs, using a filter for time.

nextToken :: Maybe Text

If the previous response was truncated, you receive this token. Use it in your next request to receive the next set of results.

sortOrder :: Maybe AutoMLSortOrder

The sort order for the results. The default is Descending.

nameContains :: Maybe Text

Request a list of jobs, using a search filter for name.

maxResults :: Maybe Natural

Request a list of jobs up to a specified limit.

creationTimeBefore :: Maybe POSIX

Request a list of jobs, using a filter for time.

lastModifiedTimeAfter :: Maybe POSIX

Request a list of jobs, using a filter for time.

sortBy :: Maybe AutoMLSortBy

The parameter by which to sort the results. The default is Name.

statusEquals :: Maybe AutoMLJobStatus

Request a list of jobs, using a filter for status.

creationTimeAfter :: Maybe POSIX

Request a list of jobs, using a filter for time.

newListAutoMLJobs :: ListAutoMLJobs

Create a value of ListAutoMLJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedTimeBefore:ListAutoMLJobs', listAutoMLJobs_lastModifiedTimeBefore - Request a list of jobs, using a filter for time.

$sel:nextToken:ListAutoMLJobs', listAutoMLJobs_nextToken - If the previous response was truncated, you receive this token. Use it in your next request to receive the next set of results.

$sel:sortOrder:ListAutoMLJobs', listAutoMLJobs_sortOrder - The sort order for the results. The default is Descending.

$sel:nameContains:ListAutoMLJobs', listAutoMLJobs_nameContains - Request a list of jobs, using a search filter for name.

$sel:maxResults:ListAutoMLJobs', listAutoMLJobs_maxResults - Request a list of jobs up to a specified limit.

$sel:creationTimeBefore:ListAutoMLJobs', listAutoMLJobs_creationTimeBefore - Request a list of jobs, using a filter for time.

$sel:lastModifiedTimeAfter:ListAutoMLJobs', listAutoMLJobs_lastModifiedTimeAfter - Request a list of jobs, using a filter for time.

$sel:sortBy:ListAutoMLJobs', listAutoMLJobs_sortBy - The parameter by which to sort the results. The default is Name.

$sel:statusEquals:ListAutoMLJobs', listAutoMLJobs_statusEquals - Request a list of jobs, using a filter for status.

$sel:creationTimeAfter:ListAutoMLJobs', listAutoMLJobs_creationTimeAfter - Request a list of jobs, using a filter for time.

Request Lenses

listAutoMLJobs_lastModifiedTimeBefore :: Lens' ListAutoMLJobs (Maybe UTCTime)

Request a list of jobs, using a filter for time.

listAutoMLJobs_nextToken :: Lens' ListAutoMLJobs (Maybe Text)

If the previous response was truncated, you receive this token. Use it in your next request to receive the next set of results.

listAutoMLJobs_sortOrder :: Lens' ListAutoMLJobs (Maybe AutoMLSortOrder)

The sort order for the results. The default is Descending.

listAutoMLJobs_nameContains :: Lens' ListAutoMLJobs (Maybe Text)

Request a list of jobs, using a search filter for name.

listAutoMLJobs_maxResults :: Lens' ListAutoMLJobs (Maybe Natural)

Request a list of jobs up to a specified limit.

listAutoMLJobs_creationTimeBefore :: Lens' ListAutoMLJobs (Maybe UTCTime)

Request a list of jobs, using a filter for time.

listAutoMLJobs_lastModifiedTimeAfter :: Lens' ListAutoMLJobs (Maybe UTCTime)

Request a list of jobs, using a filter for time.

listAutoMLJobs_sortBy :: Lens' ListAutoMLJobs (Maybe AutoMLSortBy)

The parameter by which to sort the results. The default is Name.

listAutoMLJobs_statusEquals :: Lens' ListAutoMLJobs (Maybe AutoMLJobStatus)

Request a list of jobs, using a filter for status.

listAutoMLJobs_creationTimeAfter :: Lens' ListAutoMLJobs (Maybe UTCTime)

Request a list of jobs, using a filter for time.

Destructuring the Response

data ListAutoMLJobsResponse

See: newListAutoMLJobsResponse smart constructor.

Constructors

ListAutoMLJobsResponse'

Fields

nextToken :: Maybe Text

If the previous response was truncated, you receive this token. Use it in your next request to receive the next set of results.

httpStatus :: Int

The response's http status code.

autoMLJobSummaries :: [AutoMLJobSummary]

Returns a summary list of jobs.

newListAutoMLJobsResponse

Arguments

:: Int
$sel:httpStatus:ListAutoMLJobsResponse'

-> ListAutoMLJobsResponse

Create a value of ListAutoMLJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAutoMLJobs', listAutoMLJobsResponse_nextToken - If the previous response was truncated, you receive this token. Use it in your next request to receive the next set of results.

$sel:httpStatus:ListAutoMLJobsResponse', listAutoMLJobsResponse_httpStatus - The response's http status code.

$sel:autoMLJobSummaries:ListAutoMLJobsResponse', listAutoMLJobsResponse_autoMLJobSummaries - Returns a summary list of jobs.

Response Lenses

listAutoMLJobsResponse_nextToken :: Lens' ListAutoMLJobsResponse (Maybe Text)

If the previous response was truncated, you receive this token. Use it in your next request to receive the next set of results.

listAutoMLJobsResponse_httpStatus :: Lens' ListAutoMLJobsResponse Int

The response's http status code.

listAutoMLJobsResponse_autoMLJobSummaries :: Lens' ListAutoMLJobsResponse [AutoMLJobSummary]

Returns a summary list of jobs.