Network.AWS.SageMaker.ListCandidatesForAutoMLJob

Description

List the candidates created for the job.

This operation returns paginated results.

Creating a Request

data ListCandidatesForAutoMLJob

See: newListCandidatesForAutoMLJob smart constructor.

Constructors

ListCandidatesForAutoMLJob'

Fields

nextToken :: Maybe Text

If the previous response was truncated, you receive this token. Use it in your next request to receive the next set of results.

sortOrder :: Maybe AutoMLSortOrder

The sort order for the results. The default is Ascending.

maxResults :: Maybe Natural

List the job's candidates up to a specified limit.

candidateNameEquals :: Maybe Text

List the candidates for the job and filter by candidate name.

sortBy :: Maybe CandidateSortBy

The parameter by which to sort the results. The default is Descending.

statusEquals :: Maybe CandidateStatus

List the candidates for the job and filter by status.

autoMLJobName :: Text

List the candidates created for the job by providing the job's name.

newListCandidatesForAutoMLJob

Arguments

:: Text
$sel:autoMLJobName:ListCandidatesForAutoMLJob'

-> ListCandidatesForAutoMLJob

Create a value of ListCandidatesForAutoMLJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListCandidatesForAutoMLJob', listCandidatesForAutoMLJob_nextToken - If the previous response was truncated, you receive this token. Use it in your next request to receive the next set of results.

$sel:sortOrder:ListCandidatesForAutoMLJob', listCandidatesForAutoMLJob_sortOrder - The sort order for the results. The default is Ascending.

$sel:maxResults:ListCandidatesForAutoMLJob', listCandidatesForAutoMLJob_maxResults - List the job's candidates up to a specified limit.

$sel:candidateNameEquals:ListCandidatesForAutoMLJob', listCandidatesForAutoMLJob_candidateNameEquals - List the candidates for the job and filter by candidate name.

$sel:sortBy:ListCandidatesForAutoMLJob', listCandidatesForAutoMLJob_sortBy - The parameter by which to sort the results. The default is Descending.

$sel:statusEquals:ListCandidatesForAutoMLJob', listCandidatesForAutoMLJob_statusEquals - List the candidates for the job and filter by status.

$sel:autoMLJobName:ListCandidatesForAutoMLJob', listCandidatesForAutoMLJob_autoMLJobName - List the candidates created for the job by providing the job's name.

Request Lenses

listCandidatesForAutoMLJob_nextToken :: Lens' ListCandidatesForAutoMLJob (Maybe Text)

If the previous response was truncated, you receive this token. Use it in your next request to receive the next set of results.

listCandidatesForAutoMLJob_sortOrder :: Lens' ListCandidatesForAutoMLJob (Maybe AutoMLSortOrder)

The sort order for the results. The default is Ascending.

listCandidatesForAutoMLJob_maxResults :: Lens' ListCandidatesForAutoMLJob (Maybe Natural)

List the job's candidates up to a specified limit.

listCandidatesForAutoMLJob_candidateNameEquals :: Lens' ListCandidatesForAutoMLJob (Maybe Text)

List the candidates for the job and filter by candidate name.

listCandidatesForAutoMLJob_sortBy :: Lens' ListCandidatesForAutoMLJob (Maybe CandidateSortBy)

The parameter by which to sort the results. The default is Descending.

listCandidatesForAutoMLJob_statusEquals :: Lens' ListCandidatesForAutoMLJob (Maybe CandidateStatus)

List the candidates for the job and filter by status.

listCandidatesForAutoMLJob_autoMLJobName :: Lens' ListCandidatesForAutoMLJob Text

List the candidates created for the job by providing the job's name.

Destructuring the Response

data ListCandidatesForAutoMLJobResponse

See: newListCandidatesForAutoMLJobResponse smart constructor.

Constructors

ListCandidatesForAutoMLJobResponse'

Fields

nextToken :: Maybe Text

If the previous response was truncated, you receive this token. Use it in your next request to receive the next set of results.

httpStatus :: Int

The response's http status code.

candidates :: [AutoMLCandidate]

Summaries about the AutoMLCandidates.

newListCandidatesForAutoMLJobResponse

Arguments

:: Int
$sel:httpStatus:ListCandidatesForAutoMLJobResponse'

-> ListCandidatesForAutoMLJobResponse

Create a value of ListCandidatesForAutoMLJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListCandidatesForAutoMLJob', listCandidatesForAutoMLJobResponse_nextToken - If the previous response was truncated, you receive this token. Use it in your next request to receive the next set of results.

$sel:httpStatus:ListCandidatesForAutoMLJobResponse', listCandidatesForAutoMLJobResponse_httpStatus - The response's http status code.

$sel:candidates:ListCandidatesForAutoMLJobResponse', listCandidatesForAutoMLJobResponse_candidates - Summaries about the AutoMLCandidates.

Response Lenses

listCandidatesForAutoMLJobResponse_nextToken :: Lens' ListCandidatesForAutoMLJobResponse (Maybe Text)

If the previous response was truncated, you receive this token. Use it in your next request to receive the next set of results.

listCandidatesForAutoMLJobResponse_httpStatus :: Lens' ListCandidatesForAutoMLJobResponse Int

The response's http status code.

listCandidatesForAutoMLJobResponse_candidates :: Lens' ListCandidatesForAutoMLJobResponse [AutoMLCandidate]

Summaries about the AutoMLCandidates.