Network.AWS.SageMaker.ListCodeRepositories

Description

Gets a list of the Git repositories in your account.

This operation returns paginated results.

Creating a Request

data ListCodeRepositories

See: newListCodeRepositories smart constructor.

Constructors

ListCodeRepositories'

Fields

lastModifiedTimeBefore :: Maybe POSIX

A filter that returns only Git repositories that were last modified before the specified time.

nextToken :: Maybe Text

If the result of a ListCodeRepositoriesOutput request was truncated, the response includes a NextToken. To get the next set of Git repositories, use the token in the next request.

sortOrder :: Maybe CodeRepositorySortOrder

The sort order for results. The default is Ascending.

nameContains :: Maybe Text

A string in the Git repositories name. This filter returns only repositories whose name contains the specified string.

maxResults :: Maybe Natural

The maximum number of Git repositories to return in the response.

creationTimeBefore :: Maybe POSIX

A filter that returns only Git repositories that were created before the specified time.

lastModifiedTimeAfter :: Maybe POSIX

A filter that returns only Git repositories that were last modified after the specified time.

sortBy :: Maybe CodeRepositorySortBy

The field to sort results by. The default is Name.

creationTimeAfter :: Maybe POSIX

A filter that returns only Git repositories that were created after the specified time.

newListCodeRepositories :: ListCodeRepositories

Create a value of ListCodeRepositories with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedTimeBefore:ListCodeRepositories', listCodeRepositories_lastModifiedTimeBefore - A filter that returns only Git repositories that were last modified before the specified time.

$sel:nextToken:ListCodeRepositories', listCodeRepositories_nextToken - If the result of a ListCodeRepositoriesOutput request was truncated, the response includes a NextToken. To get the next set of Git repositories, use the token in the next request.

$sel:sortOrder:ListCodeRepositories', listCodeRepositories_sortOrder - The sort order for results. The default is Ascending.

$sel:nameContains:ListCodeRepositories', listCodeRepositories_nameContains - A string in the Git repositories name. This filter returns only repositories whose name contains the specified string.

$sel:maxResults:ListCodeRepositories', listCodeRepositories_maxResults - The maximum number of Git repositories to return in the response.

$sel:creationTimeBefore:ListCodeRepositories', listCodeRepositories_creationTimeBefore - A filter that returns only Git repositories that were created before the specified time.

$sel:lastModifiedTimeAfter:ListCodeRepositories', listCodeRepositories_lastModifiedTimeAfter - A filter that returns only Git repositories that were last modified after the specified time.

$sel:sortBy:ListCodeRepositories', listCodeRepositories_sortBy - The field to sort results by. The default is Name.

$sel:creationTimeAfter:ListCodeRepositories', listCodeRepositories_creationTimeAfter - A filter that returns only Git repositories that were created after the specified time.

Request Lenses

listCodeRepositories_lastModifiedTimeBefore :: Lens' ListCodeRepositories (Maybe UTCTime)

A filter that returns only Git repositories that were last modified before the specified time.

listCodeRepositories_nextToken :: Lens' ListCodeRepositories (Maybe Text)

If the result of a ListCodeRepositoriesOutput request was truncated, the response includes a NextToken. To get the next set of Git repositories, use the token in the next request.

listCodeRepositories_sortOrder :: Lens' ListCodeRepositories (Maybe CodeRepositorySortOrder)

The sort order for results. The default is Ascending.

listCodeRepositories_nameContains :: Lens' ListCodeRepositories (Maybe Text)

A string in the Git repositories name. This filter returns only repositories whose name contains the specified string.

listCodeRepositories_maxResults :: Lens' ListCodeRepositories (Maybe Natural)

The maximum number of Git repositories to return in the response.

listCodeRepositories_creationTimeBefore :: Lens' ListCodeRepositories (Maybe UTCTime)

A filter that returns only Git repositories that were created before the specified time.

listCodeRepositories_lastModifiedTimeAfter :: Lens' ListCodeRepositories (Maybe UTCTime)

A filter that returns only Git repositories that were last modified after the specified time.

listCodeRepositories_sortBy :: Lens' ListCodeRepositories (Maybe CodeRepositorySortBy)

The field to sort results by. The default is Name.

listCodeRepositories_creationTimeAfter :: Lens' ListCodeRepositories (Maybe UTCTime)

A filter that returns only Git repositories that were created after the specified time.

Destructuring the Response

data ListCodeRepositoriesResponse

See: newListCodeRepositoriesResponse smart constructor.

Constructors

ListCodeRepositoriesResponse'

Fields

nextToken :: Maybe Text

If the result of a ListCodeRepositoriesOutput request was truncated, the response includes a NextToken. To get the next set of Git repositories, use the token in the next request.

httpStatus :: Int

The response's http status code.

codeRepositorySummaryList :: [CodeRepositorySummary]

Gets a list of summaries of the Git repositories. Each summary specifies the following values for the repository:

Name

Amazon Resource Name (ARN)

Creation time

Last modified time

Configuration information, including the URL location of the repository and the ARN of the Amazon Web Services Secrets Manager secret that contains the credentials used to access the repository.

newListCodeRepositoriesResponse

Arguments

:: Int
$sel:httpStatus:ListCodeRepositoriesResponse'

-> ListCodeRepositoriesResponse

Create a value of ListCodeRepositoriesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListCodeRepositories', listCodeRepositoriesResponse_nextToken - If the result of a ListCodeRepositoriesOutput request was truncated, the response includes a NextToken. To get the next set of Git repositories, use the token in the next request.

$sel:httpStatus:ListCodeRepositoriesResponse', listCodeRepositoriesResponse_httpStatus - The response's http status code.

$sel:codeRepositorySummaryList:ListCodeRepositoriesResponse', listCodeRepositoriesResponse_codeRepositorySummaryList - Gets a list of summaries of the Git repositories. Each summary specifies the following values for the repository:

Name

Amazon Resource Name (ARN)

Creation time

Last modified time

Configuration information, including the URL location of the repository and the ARN of the Amazon Web Services Secrets Manager secret that contains the credentials used to access the repository.

Response Lenses

listCodeRepositoriesResponse_nextToken :: Lens' ListCodeRepositoriesResponse (Maybe Text)

If the result of a ListCodeRepositoriesOutput request was truncated, the response includes a NextToken. To get the next set of Git repositories, use the token in the next request.

listCodeRepositoriesResponse_httpStatus :: Lens' ListCodeRepositoriesResponse Int

The response's http status code.

listCodeRepositoriesResponse_codeRepositorySummaryList :: Lens' ListCodeRepositoriesResponse [CodeRepositorySummary]

Gets a list of summaries of the Git repositories. Each summary specifies the following values for the repository:

Name

Amazon Resource Name (ARN)

Creation time

Last modified time

Configuration information, including the URL location of the repository and the ARN of the Amazon Web Services Secrets Manager secret that contains the credentials used to access the repository.