Network.AWS.SageMaker.ListCompilationJobs

Description

Lists model compilation jobs that satisfy various filters.

To create a model compilation job, use CreateCompilationJob. To get information about a particular model compilation job you have created, use DescribeCompilationJob.

This operation returns paginated results.

Creating a Request

data ListCompilationJobs

See: newListCompilationJobs smart constructor.

Constructors

ListCompilationJobs'

Fields

lastModifiedTimeBefore :: Maybe POSIX

A filter that returns the model compilation jobs that were modified before a specified time.

nextToken :: Maybe Text

If the result of the previous ListCompilationJobs request was truncated, the response includes a NextToken. To retrieve the next set of model compilation jobs, use the token in the next request.

sortOrder :: Maybe SortOrder

The sort order for results. The default is Ascending.

nameContains :: Maybe Text

A filter that returns the model compilation jobs whose name contains a specified string.

maxResults :: Maybe Natural

The maximum number of model compilation jobs to return in the response.

creationTimeBefore :: Maybe POSIX

A filter that returns the model compilation jobs that were created before a specified time.

lastModifiedTimeAfter :: Maybe POSIX

A filter that returns the model compilation jobs that were modified after a specified time.

sortBy :: Maybe ListCompilationJobsSortBy

The field by which to sort results. The default is CreationTime.

statusEquals :: Maybe CompilationJobStatus

A filter that retrieves model compilation jobs with a specific DescribeCompilationJobResponse$CompilationJobStatus status.

creationTimeAfter :: Maybe POSIX

A filter that returns the model compilation jobs that were created after a specified time.

newListCompilationJobs :: ListCompilationJobs

Create a value of ListCompilationJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedTimeBefore:ListCompilationJobs', listCompilationJobs_lastModifiedTimeBefore - A filter that returns the model compilation jobs that were modified before a specified time.

$sel:nextToken:ListCompilationJobs', listCompilationJobs_nextToken - If the result of the previous ListCompilationJobs request was truncated, the response includes a NextToken. To retrieve the next set of model compilation jobs, use the token in the next request.

$sel:sortOrder:ListCompilationJobs', listCompilationJobs_sortOrder - The sort order for results. The default is Ascending.

$sel:nameContains:ListCompilationJobs', listCompilationJobs_nameContains - A filter that returns the model compilation jobs whose name contains a specified string.

$sel:maxResults:ListCompilationJobs', listCompilationJobs_maxResults - The maximum number of model compilation jobs to return in the response.

$sel:creationTimeBefore:ListCompilationJobs', listCompilationJobs_creationTimeBefore - A filter that returns the model compilation jobs that were created before a specified time.

$sel:lastModifiedTimeAfter:ListCompilationJobs', listCompilationJobs_lastModifiedTimeAfter - A filter that returns the model compilation jobs that were modified after a specified time.

$sel:sortBy:ListCompilationJobs', listCompilationJobs_sortBy - The field by which to sort results. The default is CreationTime.

$sel:statusEquals:ListCompilationJobs', listCompilationJobs_statusEquals - A filter that retrieves model compilation jobs with a specific DescribeCompilationJobResponse$CompilationJobStatus status.

$sel:creationTimeAfter:ListCompilationJobs', listCompilationJobs_creationTimeAfter - A filter that returns the model compilation jobs that were created after a specified time.

Request Lenses

listCompilationJobs_lastModifiedTimeBefore :: Lens' ListCompilationJobs (Maybe UTCTime)

A filter that returns the model compilation jobs that were modified before a specified time.

listCompilationJobs_nextToken :: Lens' ListCompilationJobs (Maybe Text)

If the result of the previous ListCompilationJobs request was truncated, the response includes a NextToken. To retrieve the next set of model compilation jobs, use the token in the next request.

listCompilationJobs_sortOrder :: Lens' ListCompilationJobs (Maybe SortOrder)

The sort order for results. The default is Ascending.

listCompilationJobs_nameContains :: Lens' ListCompilationJobs (Maybe Text)

A filter that returns the model compilation jobs whose name contains a specified string.

listCompilationJobs_maxResults :: Lens' ListCompilationJobs (Maybe Natural)

The maximum number of model compilation jobs to return in the response.

listCompilationJobs_creationTimeBefore :: Lens' ListCompilationJobs (Maybe UTCTime)

A filter that returns the model compilation jobs that were created before a specified time.

listCompilationJobs_lastModifiedTimeAfter :: Lens' ListCompilationJobs (Maybe UTCTime)

A filter that returns the model compilation jobs that were modified after a specified time.

listCompilationJobs_sortBy :: Lens' ListCompilationJobs (Maybe ListCompilationJobsSortBy)

The field by which to sort results. The default is CreationTime.

listCompilationJobs_statusEquals :: Lens' ListCompilationJobs (Maybe CompilationJobStatus)

A filter that retrieves model compilation jobs with a specific DescribeCompilationJobResponse$CompilationJobStatus status.

listCompilationJobs_creationTimeAfter :: Lens' ListCompilationJobs (Maybe UTCTime)

A filter that returns the model compilation jobs that were created after a specified time.

Destructuring the Response

data ListCompilationJobsResponse

See: newListCompilationJobsResponse smart constructor.

Constructors

ListCompilationJobsResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, Amazon SageMaker returns this NextToken. To retrieve the next set of model compilation jobs, use this token in the next request.

httpStatus :: Int

The response's http status code.

compilationJobSummaries :: [CompilationJobSummary]

An array of CompilationJobSummary objects, each describing a model compilation job.

newListCompilationJobsResponse

Arguments

:: Int
$sel:httpStatus:ListCompilationJobsResponse'

-> ListCompilationJobsResponse

Create a value of ListCompilationJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListCompilationJobs', listCompilationJobsResponse_nextToken - If the response is truncated, Amazon SageMaker returns this NextToken. To retrieve the next set of model compilation jobs, use this token in the next request.

$sel:httpStatus:ListCompilationJobsResponse', listCompilationJobsResponse_httpStatus - The response's http status code.

$sel:compilationJobSummaries:ListCompilationJobsResponse', listCompilationJobsResponse_compilationJobSummaries - An array of CompilationJobSummary objects, each describing a model compilation job.

Response Lenses

listCompilationJobsResponse_nextToken :: Lens' ListCompilationJobsResponse (Maybe Text)

If the response is truncated, Amazon SageMaker returns this NextToken. To retrieve the next set of model compilation jobs, use this token in the next request.

listCompilationJobsResponse_httpStatus :: Lens' ListCompilationJobsResponse Int

The response's http status code.

listCompilationJobsResponse_compilationJobSummaries :: Lens' ListCompilationJobsResponse [CompilationJobSummary]

An array of CompilationJobSummary objects, each describing a model compilation job.