Network.AWS.SageMaker.ListContexts

Description

Lists the contexts in your account and their properties.

This operation returns paginated results.

Creating a Request

data ListContexts

See: newListContexts smart constructor.

Constructors

ListContexts'

Fields

contextType :: Maybe Text

A filter that returns only contexts of the specified type.

nextToken :: Maybe Text

If the previous call to ListContexts didn't return the full set of contexts, the call returns a token for getting the next set of contexts.

sortOrder :: Maybe SortOrder

The sort order. The default value is Descending.

createdAfter :: Maybe POSIX

A filter that returns only contexts created on or after the specified time.

createdBefore :: Maybe POSIX

A filter that returns only contexts created on or before the specified time.

maxResults :: Maybe Natural

The maximum number of contexts to return in the response. The default value is 10.

sourceUri :: Maybe Text

A filter that returns only contexts with the specified source URI.

sortBy :: Maybe SortContextsBy

The property used to sort results. The default value is CreationTime.

newListContexts :: ListContexts

Create a value of ListContexts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contextType:ListContexts', listContexts_contextType - A filter that returns only contexts of the specified type.

$sel:nextToken:ListContexts', listContexts_nextToken - If the previous call to ListContexts didn't return the full set of contexts, the call returns a token for getting the next set of contexts.

$sel:sortOrder:ListContexts', listContexts_sortOrder - The sort order. The default value is Descending.

$sel:createdAfter:ListContexts', listContexts_createdAfter - A filter that returns only contexts created on or after the specified time.

$sel:createdBefore:ListContexts', listContexts_createdBefore - A filter that returns only contexts created on or before the specified time.

$sel:maxResults:ListContexts', listContexts_maxResults - The maximum number of contexts to return in the response. The default value is 10.

$sel:sourceUri:ListContexts', listContexts_sourceUri - A filter that returns only contexts with the specified source URI.

$sel:sortBy:ListContexts', listContexts_sortBy - The property used to sort results. The default value is CreationTime.

Request Lenses

listContexts_contextType :: Lens' ListContexts (Maybe Text)

A filter that returns only contexts of the specified type.

listContexts_nextToken :: Lens' ListContexts (Maybe Text)

If the previous call to ListContexts didn't return the full set of contexts, the call returns a token for getting the next set of contexts.

listContexts_sortOrder :: Lens' ListContexts (Maybe SortOrder)

The sort order. The default value is Descending.

listContexts_createdAfter :: Lens' ListContexts (Maybe UTCTime)

A filter that returns only contexts created on or after the specified time.

listContexts_createdBefore :: Lens' ListContexts (Maybe UTCTime)

A filter that returns only contexts created on or before the specified time.

listContexts_maxResults :: Lens' ListContexts (Maybe Natural)

The maximum number of contexts to return in the response. The default value is 10.

listContexts_sourceUri :: Lens' ListContexts (Maybe Text)

A filter that returns only contexts with the specified source URI.

listContexts_sortBy :: Lens' ListContexts (Maybe SortContextsBy)

The property used to sort results. The default value is CreationTime.

Destructuring the Response

data ListContextsResponse

See: newListContextsResponse smart constructor.

Constructors

ListContextsResponse'

Fields

nextToken :: Maybe Text

A token for getting the next set of contexts, if there are any.

contextSummaries :: Maybe [ContextSummary]

A list of contexts and their properties.

httpStatus :: Int

The response's http status code.

newListContextsResponse

Arguments

:: Int
$sel:httpStatus:ListContextsResponse'

-> ListContextsResponse

Create a value of ListContextsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListContexts', listContextsResponse_nextToken - A token for getting the next set of contexts, if there are any.

$sel:contextSummaries:ListContextsResponse', listContextsResponse_contextSummaries - A list of contexts and their properties.

$sel:httpStatus:ListContextsResponse', listContextsResponse_httpStatus - The response's http status code.

Response Lenses

listContextsResponse_nextToken :: Lens' ListContextsResponse (Maybe Text)

A token for getting the next set of contexts, if there are any.

listContextsResponse_contextSummaries :: Lens' ListContextsResponse (Maybe [ContextSummary])

A list of contexts and their properties.

listContextsResponse_httpStatus :: Lens' ListContextsResponse Int

The response's http status code.