Network.AWS.SageMaker.ListDataQualityJobDefinitions

Description

Lists the data quality job definitions in your account.

This operation returns paginated results.

Creating a Request

data ListDataQualityJobDefinitions

See: newListDataQualityJobDefinitions smart constructor.

Constructors

ListDataQualityJobDefinitions'

Fields

nextToken :: Maybe Text

If the result of the previous ListDataQualityJobDefinitions request was truncated, the response includes a NextToken. To retrieve the next set of transform jobs, use the token in the next request.>

sortOrder :: Maybe SortOrder

The sort order for results. The default is Descending.

endpointName :: Maybe Text

A filter that lists the data quality job definitions associated with the specified endpoint.

nameContains :: Maybe Text

A string in the data quality monitoring job definition name. This filter returns only data quality monitoring job definitions whose name contains the specified string.

maxResults :: Maybe Natural

The maximum number of data quality monitoring job definitions to return in the response.

creationTimeBefore :: Maybe POSIX

A filter that returns only data quality monitoring job definitions created before the specified time.

sortBy :: Maybe MonitoringJobDefinitionSortKey

The field to sort results by. The default is CreationTime.

creationTimeAfter :: Maybe POSIX

A filter that returns only data quality monitoring job definitions created after the specified time.

newListDataQualityJobDefinitions :: ListDataQualityJobDefinitions

Create a value of ListDataQualityJobDefinitions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDataQualityJobDefinitions', listDataQualityJobDefinitions_nextToken - If the result of the previous ListDataQualityJobDefinitions request was truncated, the response includes a NextToken. To retrieve the next set of transform jobs, use the token in the next request.>

$sel:sortOrder:ListDataQualityJobDefinitions', listDataQualityJobDefinitions_sortOrder - The sort order for results. The default is Descending.

$sel:endpointName:ListDataQualityJobDefinitions', listDataQualityJobDefinitions_endpointName - A filter that lists the data quality job definitions associated with the specified endpoint.

$sel:nameContains:ListDataQualityJobDefinitions', listDataQualityJobDefinitions_nameContains - A string in the data quality monitoring job definition name. This filter returns only data quality monitoring job definitions whose name contains the specified string.

$sel:maxResults:ListDataQualityJobDefinitions', listDataQualityJobDefinitions_maxResults - The maximum number of data quality monitoring job definitions to return in the response.

$sel:creationTimeBefore:ListDataQualityJobDefinitions', listDataQualityJobDefinitions_creationTimeBefore - A filter that returns only data quality monitoring job definitions created before the specified time.

$sel:sortBy:ListDataQualityJobDefinitions', listDataQualityJobDefinitions_sortBy - The field to sort results by. The default is CreationTime.

$sel:creationTimeAfter:ListDataQualityJobDefinitions', listDataQualityJobDefinitions_creationTimeAfter - A filter that returns only data quality monitoring job definitions created after the specified time.

Request Lenses

listDataQualityJobDefinitions_nextToken :: Lens' ListDataQualityJobDefinitions (Maybe Text)

If the result of the previous ListDataQualityJobDefinitions request was truncated, the response includes a NextToken. To retrieve the next set of transform jobs, use the token in the next request.>

listDataQualityJobDefinitions_sortOrder :: Lens' ListDataQualityJobDefinitions (Maybe SortOrder)

The sort order for results. The default is Descending.

listDataQualityJobDefinitions_endpointName :: Lens' ListDataQualityJobDefinitions (Maybe Text)

A filter that lists the data quality job definitions associated with the specified endpoint.

listDataQualityJobDefinitions_nameContains :: Lens' ListDataQualityJobDefinitions (Maybe Text)

A string in the data quality monitoring job definition name. This filter returns only data quality monitoring job definitions whose name contains the specified string.

listDataQualityJobDefinitions_maxResults :: Lens' ListDataQualityJobDefinitions (Maybe Natural)

The maximum number of data quality monitoring job definitions to return in the response.

listDataQualityJobDefinitions_creationTimeBefore :: Lens' ListDataQualityJobDefinitions (Maybe UTCTime)

A filter that returns only data quality monitoring job definitions created before the specified time.

listDataQualityJobDefinitions_sortBy :: Lens' ListDataQualityJobDefinitions (Maybe MonitoringJobDefinitionSortKey)

The field to sort results by. The default is CreationTime.

listDataQualityJobDefinitions_creationTimeAfter :: Lens' ListDataQualityJobDefinitions (Maybe UTCTime)

A filter that returns only data quality monitoring job definitions created after the specified time.

Destructuring the Response

data ListDataQualityJobDefinitionsResponse

See: newListDataQualityJobDefinitionsResponse smart constructor.

Constructors

ListDataQualityJobDefinitionsResponse'

Fields

nextToken :: Maybe Text

If the result of the previous ListDataQualityJobDefinitions request was truncated, the response includes a NextToken. To retrieve the next set of data quality monitoring job definitions, use the token in the next request.

httpStatus :: Int

The response's http status code.

jobDefinitionSummaries :: [MonitoringJobDefinitionSummary]

A list of data quality monitoring job definitions.

newListDataQualityJobDefinitionsResponse

Arguments

:: Int
$sel:httpStatus:ListDataQualityJobDefinitionsResponse'

-> ListDataQualityJobDefinitionsResponse

Create a value of ListDataQualityJobDefinitionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDataQualityJobDefinitions', listDataQualityJobDefinitionsResponse_nextToken - If the result of the previous ListDataQualityJobDefinitions request was truncated, the response includes a NextToken. To retrieve the next set of data quality monitoring job definitions, use the token in the next request.

$sel:httpStatus:ListDataQualityJobDefinitionsResponse', listDataQualityJobDefinitionsResponse_httpStatus - The response's http status code.

$sel:jobDefinitionSummaries:ListDataQualityJobDefinitionsResponse', listDataQualityJobDefinitionsResponse_jobDefinitionSummaries - A list of data quality monitoring job definitions.

Response Lenses

listDataQualityJobDefinitionsResponse_nextToken :: Lens' ListDataQualityJobDefinitionsResponse (Maybe Text)

If the result of the previous ListDataQualityJobDefinitions request was truncated, the response includes a NextToken. To retrieve the next set of data quality monitoring job definitions, use the token in the next request.

listDataQualityJobDefinitionsResponse_httpStatus :: Lens' ListDataQualityJobDefinitionsResponse Int

The response's http status code.

listDataQualityJobDefinitionsResponse_jobDefinitionSummaries :: Lens' ListDataQualityJobDefinitionsResponse [MonitoringJobDefinitionSummary]

A list of data quality monitoring job definitions.