Network.AWS.SageMaker.ListDeviceFleets

Description

Returns a list of devices in the fleet.

This operation returns paginated results.

Creating a Request

data ListDeviceFleets

See: newListDeviceFleets smart constructor.

Constructors

ListDeviceFleets'

Fields

lastModifiedTimeBefore :: Maybe POSIX

Select fleets where the job was updated before X

nextToken :: Maybe Text

The response from the last list when returning a list large enough to need tokening.

sortOrder :: Maybe SortOrder

What direction to sort in.

nameContains :: Maybe Text

Filter for fleets containing this name in their fleet device name.

maxResults :: Maybe Int

The maximum number of results to select.

creationTimeBefore :: Maybe POSIX

Filter fleets where the edge packaging job was created before specified time.

lastModifiedTimeAfter :: Maybe POSIX

Select fleets where the job was updated after X

sortBy :: Maybe ListDeviceFleetsSortBy

The column to sort by.

creationTimeAfter :: Maybe POSIX

Filter fleets where packaging job was created after specified time.

newListDeviceFleets :: ListDeviceFleets

Create a value of ListDeviceFleets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedTimeBefore:ListDeviceFleets', listDeviceFleets_lastModifiedTimeBefore - Select fleets where the job was updated before X

$sel:nextToken:ListDeviceFleets', listDeviceFleets_nextToken - The response from the last list when returning a list large enough to need tokening.

$sel:sortOrder:ListDeviceFleets', listDeviceFleets_sortOrder - What direction to sort in.

$sel:nameContains:ListDeviceFleets', listDeviceFleets_nameContains - Filter for fleets containing this name in their fleet device name.

$sel:maxResults:ListDeviceFleets', listDeviceFleets_maxResults - The maximum number of results to select.

$sel:creationTimeBefore:ListDeviceFleets', listDeviceFleets_creationTimeBefore - Filter fleets where the edge packaging job was created before specified time.

$sel:lastModifiedTimeAfter:ListDeviceFleets', listDeviceFleets_lastModifiedTimeAfter - Select fleets where the job was updated after X

$sel:sortBy:ListDeviceFleets', listDeviceFleets_sortBy - The column to sort by.

$sel:creationTimeAfter:ListDeviceFleets', listDeviceFleets_creationTimeAfter - Filter fleets where packaging job was created after specified time.

Request Lenses

listDeviceFleets_lastModifiedTimeBefore :: Lens' ListDeviceFleets (Maybe UTCTime)

Select fleets where the job was updated before X

listDeviceFleets_nextToken :: Lens' ListDeviceFleets (Maybe Text)

The response from the last list when returning a list large enough to need tokening.

listDeviceFleets_sortOrder :: Lens' ListDeviceFleets (Maybe SortOrder)

What direction to sort in.

listDeviceFleets_nameContains :: Lens' ListDeviceFleets (Maybe Text)

Filter for fleets containing this name in their fleet device name.

listDeviceFleets_maxResults :: Lens' ListDeviceFleets (Maybe Int)

The maximum number of results to select.

listDeviceFleets_creationTimeBefore :: Lens' ListDeviceFleets (Maybe UTCTime)

Filter fleets where the edge packaging job was created before specified time.

listDeviceFleets_lastModifiedTimeAfter :: Lens' ListDeviceFleets (Maybe UTCTime)

Select fleets where the job was updated after X

listDeviceFleets_sortBy :: Lens' ListDeviceFleets (Maybe ListDeviceFleetsSortBy)

The column to sort by.

listDeviceFleets_creationTimeAfter :: Lens' ListDeviceFleets (Maybe UTCTime)

Filter fleets where packaging job was created after specified time.

Destructuring the Response

data ListDeviceFleetsResponse

See: newListDeviceFleetsResponse smart constructor.

Constructors

ListDeviceFleetsResponse'

Fields

nextToken :: Maybe Text

The response from the last list when returning a list large enough to need tokening.

httpStatus :: Int

The response's http status code.

deviceFleetSummaries :: [DeviceFleetSummary]

Summary of the device fleet.

newListDeviceFleetsResponse

Arguments

:: Int
$sel:httpStatus:ListDeviceFleetsResponse'

-> ListDeviceFleetsResponse

Create a value of ListDeviceFleetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDeviceFleets', listDeviceFleetsResponse_nextToken - The response from the last list when returning a list large enough to need tokening.

$sel:httpStatus:ListDeviceFleetsResponse', listDeviceFleetsResponse_httpStatus - The response's http status code.

$sel:deviceFleetSummaries:ListDeviceFleetsResponse', listDeviceFleetsResponse_deviceFleetSummaries - Summary of the device fleet.

Response Lenses

listDeviceFleetsResponse_nextToken :: Lens' ListDeviceFleetsResponse (Maybe Text)

The response from the last list when returning a list large enough to need tokening.

listDeviceFleetsResponse_httpStatus :: Lens' ListDeviceFleetsResponse Int

The response's http status code.

listDeviceFleetsResponse_deviceFleetSummaries :: Lens' ListDeviceFleetsResponse [DeviceFleetSummary]

Summary of the device fleet.