Network.AWS.SageMaker.ListEndpointConfigs

Description

Lists endpoint configurations.

This operation returns paginated results.

Creating a Request

data ListEndpointConfigs

See: newListEndpointConfigs smart constructor.

Constructors

ListEndpointConfigs'

Fields

nameContains :: Maybe Text

A string in the endpoint configuration name. This filter returns only endpoint configurations whose name contains the specified string.

creationTimeAfter :: Maybe POSIX

A filter that returns only endpoint configurations with a creation time greater than or equal to the specified time (timestamp).

nextToken :: Maybe Text

If the result of the previous ListEndpointConfig request was truncated, the response includes a NextToken. To retrieve the next set of endpoint configurations, use the token in the next request.

sortOrder :: Maybe OrderKey

The sort order for results. The default is Descending.

creationTimeBefore :: Maybe POSIX

A filter that returns only endpoint configurations created before the specified time (timestamp).

maxResults :: Maybe Natural

The maximum number of training jobs to return in the response.

sortBy :: Maybe EndpointConfigSortKey

The field to sort results by. The default is CreationTime.

newListEndpointConfigs :: ListEndpointConfigs

Create a value of ListEndpointConfigs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nameContains:ListEndpointConfigs', listEndpointConfigs_nameContains - A string in the endpoint configuration name. This filter returns only endpoint configurations whose name contains the specified string.

$sel:creationTimeAfter:ListEndpointConfigs', listEndpointConfigs_creationTimeAfter - A filter that returns only endpoint configurations with a creation time greater than or equal to the specified time (timestamp).

$sel:nextToken:ListEndpointConfigs', listEndpointConfigs_nextToken - If the result of the previous ListEndpointConfig request was truncated, the response includes a NextToken. To retrieve the next set of endpoint configurations, use the token in the next request.

$sel:sortOrder:ListEndpointConfigs', listEndpointConfigs_sortOrder - The sort order for results. The default is Descending.

$sel:creationTimeBefore:ListEndpointConfigs', listEndpointConfigs_creationTimeBefore - A filter that returns only endpoint configurations created before the specified time (timestamp).

$sel:maxResults:ListEndpointConfigs', listEndpointConfigs_maxResults - The maximum number of training jobs to return in the response.

$sel:sortBy:ListEndpointConfigs', listEndpointConfigs_sortBy - The field to sort results by. The default is CreationTime.

Request Lenses

listEndpointConfigs_nameContains :: Lens' ListEndpointConfigs (Maybe Text)

A string in the endpoint configuration name. This filter returns only endpoint configurations whose name contains the specified string.

listEndpointConfigs_creationTimeAfter :: Lens' ListEndpointConfigs (Maybe UTCTime)

A filter that returns only endpoint configurations with a creation time greater than or equal to the specified time (timestamp).

listEndpointConfigs_nextToken :: Lens' ListEndpointConfigs (Maybe Text)

If the result of the previous ListEndpointConfig request was truncated, the response includes a NextToken. To retrieve the next set of endpoint configurations, use the token in the next request.

listEndpointConfigs_sortOrder :: Lens' ListEndpointConfigs (Maybe OrderKey)

The sort order for results. The default is Descending.

listEndpointConfigs_creationTimeBefore :: Lens' ListEndpointConfigs (Maybe UTCTime)

A filter that returns only endpoint configurations created before the specified time (timestamp).

listEndpointConfigs_maxResults :: Lens' ListEndpointConfigs (Maybe Natural)

The maximum number of training jobs to return in the response.

listEndpointConfigs_sortBy :: Lens' ListEndpointConfigs (Maybe EndpointConfigSortKey)

The field to sort results by. The default is CreationTime.

Destructuring the Response

data ListEndpointConfigsResponse

See: newListEndpointConfigsResponse smart constructor.

Constructors

ListEndpointConfigsResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of endpoint configurations, use it in the subsequent request

httpStatus :: Int

The response's http status code.

endpointConfigs :: [EndpointConfigSummary]

An array of endpoint configurations.

newListEndpointConfigsResponse

Arguments

:: Int
$sel:httpStatus:ListEndpointConfigsResponse'

-> ListEndpointConfigsResponse

Create a value of ListEndpointConfigsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEndpointConfigs', listEndpointConfigsResponse_nextToken - If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of endpoint configurations, use it in the subsequent request

$sel:httpStatus:ListEndpointConfigsResponse', listEndpointConfigsResponse_httpStatus - The response's http status code.

$sel:endpointConfigs:ListEndpointConfigsResponse', listEndpointConfigsResponse_endpointConfigs - An array of endpoint configurations.

Response Lenses

listEndpointConfigsResponse_nextToken :: Lens' ListEndpointConfigsResponse (Maybe Text)

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of endpoint configurations, use it in the subsequent request

listEndpointConfigsResponse_httpStatus :: Lens' ListEndpointConfigsResponse Int

The response's http status code.

listEndpointConfigsResponse_endpointConfigs :: Lens' ListEndpointConfigsResponse [EndpointConfigSummary]

An array of endpoint configurations.