Network.AWS.SageMaker.ListExperiments

Description

Lists all the experiments in your account. The list can be filtered to show only experiments that were created in a specific time range. The list can be sorted by experiment name or creation time.

This operation returns paginated results.

Creating a Request

data ListExperiments

See: newListExperiments smart constructor.

Constructors

ListExperiments'

Fields

createdAfter :: Maybe POSIX

A filter that returns only experiments created after the specified time.

nextToken :: Maybe Text

If the previous call to ListExperiments didn't return the full set of experiments, the call returns a token for getting the next set of experiments.

sortOrder :: Maybe SortOrder

The sort order. The default value is Descending.

maxResults :: Maybe Natural

The maximum number of experiments to return in the response. The default value is 10.

createdBefore :: Maybe POSIX

A filter that returns only experiments created before the specified time.

sortBy :: Maybe SortExperimentsBy

The property used to sort results. The default value is CreationTime.

newListExperiments :: ListExperiments

Create a value of ListExperiments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAfter:ListExperiments', listExperiments_createdAfter - A filter that returns only experiments created after the specified time.

$sel:nextToken:ListExperiments', listExperiments_nextToken - If the previous call to ListExperiments didn't return the full set of experiments, the call returns a token for getting the next set of experiments.

$sel:sortOrder:ListExperiments', listExperiments_sortOrder - The sort order. The default value is Descending.

$sel:maxResults:ListExperiments', listExperiments_maxResults - The maximum number of experiments to return in the response. The default value is 10.

$sel:createdBefore:ListExperiments', listExperiments_createdBefore - A filter that returns only experiments created before the specified time.

$sel:sortBy:ListExperiments', listExperiments_sortBy - The property used to sort results. The default value is CreationTime.

Request Lenses

listExperiments_createdAfter :: Lens' ListExperiments (Maybe UTCTime)

A filter that returns only experiments created after the specified time.

listExperiments_nextToken :: Lens' ListExperiments (Maybe Text)

If the previous call to ListExperiments didn't return the full set of experiments, the call returns a token for getting the next set of experiments.

listExperiments_sortOrder :: Lens' ListExperiments (Maybe SortOrder)

The sort order. The default value is Descending.

listExperiments_maxResults :: Lens' ListExperiments (Maybe Natural)

The maximum number of experiments to return in the response. The default value is 10.

listExperiments_createdBefore :: Lens' ListExperiments (Maybe UTCTime)

A filter that returns only experiments created before the specified time.

listExperiments_sortBy :: Lens' ListExperiments (Maybe SortExperimentsBy)

The property used to sort results. The default value is CreationTime.

Destructuring the Response

data ListExperimentsResponse

See: newListExperimentsResponse smart constructor.

Constructors

ListExperimentsResponse'

Fields

experimentSummaries :: Maybe [ExperimentSummary]

A list of the summaries of your experiments.

nextToken :: Maybe Text

A token for getting the next set of experiments, if there are any.

httpStatus :: Int

The response's http status code.

newListExperimentsResponse

Arguments

:: Int
$sel:httpStatus:ListExperimentsResponse'

-> ListExperimentsResponse

Create a value of ListExperimentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:experimentSummaries:ListExperimentsResponse', listExperimentsResponse_experimentSummaries - A list of the summaries of your experiments.

$sel:nextToken:ListExperiments', listExperimentsResponse_nextToken - A token for getting the next set of experiments, if there are any.

$sel:httpStatus:ListExperimentsResponse', listExperimentsResponse_httpStatus - The response's http status code.

Response Lenses

listExperimentsResponse_experimentSummaries :: Lens' ListExperimentsResponse (Maybe [ExperimentSummary])

A list of the summaries of your experiments.

listExperimentsResponse_nextToken :: Lens' ListExperimentsResponse (Maybe Text)

A token for getting the next set of experiments, if there are any.

listExperimentsResponse_httpStatus :: Lens' ListExperimentsResponse Int

The response's http status code.