Network.AWS.SageMaker.ListFeatureGroups

Description

List FeatureGroups based on given filter and order.

This operation returns paginated results.

Creating a Request

data ListFeatureGroups

See: newListFeatureGroups smart constructor.

Constructors

ListFeatureGroups'

Fields

nextToken :: Maybe Text

A token to resume pagination of ListFeatureGroups results.

sortOrder :: Maybe FeatureGroupSortOrder

The order in which feature groups are listed.

nameContains :: Maybe Text

A string that partially matches one or more FeatureGroups names. Filters FeatureGroups by name.

maxResults :: Maybe Natural

The maximum number of results returned by ListFeatureGroups.

creationTimeBefore :: Maybe POSIX

Use this parameter to search for FeatureGroupss created before a specific date and time.

offlineStoreStatusEquals :: Maybe OfflineStoreStatusValue

An OfflineStore status. Filters by OfflineStore status.

sortBy :: Maybe FeatureGroupSortBy

The value on which the feature group list is sorted.

featureGroupStatusEquals :: Maybe FeatureGroupStatus

A FeatureGroup status. Filters by FeatureGroup status.

creationTimeAfter :: Maybe POSIX

Use this parameter to search for FeatureGroupss created after a specific date and time.

newListFeatureGroups :: ListFeatureGroups

Create a value of ListFeatureGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFeatureGroups', listFeatureGroups_nextToken - A token to resume pagination of ListFeatureGroups results.

$sel:sortOrder:ListFeatureGroups', listFeatureGroups_sortOrder - The order in which feature groups are listed.

$sel:nameContains:ListFeatureGroups', listFeatureGroups_nameContains - A string that partially matches one or more FeatureGroups names. Filters FeatureGroups by name.

$sel:maxResults:ListFeatureGroups', listFeatureGroups_maxResults - The maximum number of results returned by ListFeatureGroups.

$sel:creationTimeBefore:ListFeatureGroups', listFeatureGroups_creationTimeBefore - Use this parameter to search for FeatureGroupss created before a specific date and time.

$sel:offlineStoreStatusEquals:ListFeatureGroups', listFeatureGroups_offlineStoreStatusEquals - An OfflineStore status. Filters by OfflineStore status.

$sel:sortBy:ListFeatureGroups', listFeatureGroups_sortBy - The value on which the feature group list is sorted.

$sel:featureGroupStatusEquals:ListFeatureGroups', listFeatureGroups_featureGroupStatusEquals - A FeatureGroup status. Filters by FeatureGroup status.

$sel:creationTimeAfter:ListFeatureGroups', listFeatureGroups_creationTimeAfter - Use this parameter to search for FeatureGroupss created after a specific date and time.

Request Lenses

listFeatureGroups_nextToken :: Lens' ListFeatureGroups (Maybe Text)

A token to resume pagination of ListFeatureGroups results.

listFeatureGroups_sortOrder :: Lens' ListFeatureGroups (Maybe FeatureGroupSortOrder)

The order in which feature groups are listed.

listFeatureGroups_nameContains :: Lens' ListFeatureGroups (Maybe Text)

A string that partially matches one or more FeatureGroups names. Filters FeatureGroups by name.

listFeatureGroups_maxResults :: Lens' ListFeatureGroups (Maybe Natural)

The maximum number of results returned by ListFeatureGroups.

listFeatureGroups_creationTimeBefore :: Lens' ListFeatureGroups (Maybe UTCTime)

Use this parameter to search for FeatureGroupss created before a specific date and time.

listFeatureGroups_offlineStoreStatusEquals :: Lens' ListFeatureGroups (Maybe OfflineStoreStatusValue)

An OfflineStore status. Filters by OfflineStore status.

listFeatureGroups_sortBy :: Lens' ListFeatureGroups (Maybe FeatureGroupSortBy)

The value on which the feature group list is sorted.

listFeatureGroups_featureGroupStatusEquals :: Lens' ListFeatureGroups (Maybe FeatureGroupStatus)

A FeatureGroup status. Filters by FeatureGroup status.

listFeatureGroups_creationTimeAfter :: Lens' ListFeatureGroups (Maybe UTCTime)

Use this parameter to search for FeatureGroupss created after a specific date and time.

Destructuring the Response

data ListFeatureGroupsResponse

See: newListFeatureGroupsResponse smart constructor.

Constructors

ListFeatureGroupsResponse'

Fields

nextToken :: Maybe Text

A token to resume pagination of ListFeatureGroups results.

httpStatus :: Int

The response's http status code.

featureGroupSummaries :: [FeatureGroupSummary]

A summary of feature groups.

newListFeatureGroupsResponse

Arguments

:: Int
$sel:httpStatus:ListFeatureGroupsResponse'

-> ListFeatureGroupsResponse

Create a value of ListFeatureGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFeatureGroups', listFeatureGroupsResponse_nextToken - A token to resume pagination of ListFeatureGroups results.

$sel:httpStatus:ListFeatureGroupsResponse', listFeatureGroupsResponse_httpStatus - The response's http status code.

$sel:featureGroupSummaries:ListFeatureGroupsResponse', listFeatureGroupsResponse_featureGroupSummaries - A summary of feature groups.

Response Lenses

listFeatureGroupsResponse_nextToken :: Lens' ListFeatureGroupsResponse (Maybe Text)

A token to resume pagination of ListFeatureGroups results.

listFeatureGroupsResponse_httpStatus :: Lens' ListFeatureGroupsResponse Int

The response's http status code.

listFeatureGroupsResponse_featureGroupSummaries :: Lens' ListFeatureGroupsResponse [FeatureGroupSummary]

A summary of feature groups.