Network.AWS.SageMaker.ListFlowDefinitions

Description

Returns information about the flow definitions in your account.

This operation returns paginated results.

Creating a Request

data ListFlowDefinitions

See: newListFlowDefinitions smart constructor.

Constructors

ListFlowDefinitions'

Fields

nextToken :: Maybe Text

A token to resume pagination.

sortOrder :: Maybe SortOrder

An optional value that specifies whether you want the results sorted in Ascending or Descending order.

maxResults :: Maybe Natural

The total number of items to return. If the total number of available items is more than the value specified in MaxResults, then a NextToken will be provided in the output that you can use to resume pagination.

creationTimeBefore :: Maybe POSIX

A filter that returns only flow definitions that were created before the specified timestamp.

creationTimeAfter :: Maybe POSIX

A filter that returns only flow definitions with a creation time greater than or equal to the specified timestamp.

newListFlowDefinitions :: ListFlowDefinitions

Create a value of ListFlowDefinitions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFlowDefinitions', listFlowDefinitions_nextToken - A token to resume pagination.

$sel:sortOrder:ListFlowDefinitions', listFlowDefinitions_sortOrder - An optional value that specifies whether you want the results sorted in Ascending or Descending order.

$sel:maxResults:ListFlowDefinitions', listFlowDefinitions_maxResults - The total number of items to return. If the total number of available items is more than the value specified in MaxResults, then a NextToken will be provided in the output that you can use to resume pagination.

$sel:creationTimeBefore:ListFlowDefinitions', listFlowDefinitions_creationTimeBefore - A filter that returns only flow definitions that were created before the specified timestamp.

$sel:creationTimeAfter:ListFlowDefinitions', listFlowDefinitions_creationTimeAfter - A filter that returns only flow definitions with a creation time greater than or equal to the specified timestamp.

Request Lenses

listFlowDefinitions_nextToken :: Lens' ListFlowDefinitions (Maybe Text)

A token to resume pagination.

listFlowDefinitions_sortOrder :: Lens' ListFlowDefinitions (Maybe SortOrder)

An optional value that specifies whether you want the results sorted in Ascending or Descending order.

listFlowDefinitions_maxResults :: Lens' ListFlowDefinitions (Maybe Natural)

The total number of items to return. If the total number of available items is more than the value specified in MaxResults, then a NextToken will be provided in the output that you can use to resume pagination.

listFlowDefinitions_creationTimeBefore :: Lens' ListFlowDefinitions (Maybe UTCTime)

A filter that returns only flow definitions that were created before the specified timestamp.

listFlowDefinitions_creationTimeAfter :: Lens' ListFlowDefinitions (Maybe UTCTime)

A filter that returns only flow definitions with a creation time greater than or equal to the specified timestamp.

Destructuring the Response

data ListFlowDefinitionsResponse

See: newListFlowDefinitionsResponse smart constructor.

Constructors

ListFlowDefinitionsResponse'

Fields

nextToken :: Maybe Text

A token to resume pagination.

httpStatus :: Int

The response's http status code.

flowDefinitionSummaries :: [FlowDefinitionSummary]

An array of objects describing the flow definitions.

newListFlowDefinitionsResponse

Arguments

:: Int
$sel:httpStatus:ListFlowDefinitionsResponse'

-> ListFlowDefinitionsResponse

Create a value of ListFlowDefinitionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFlowDefinitions', listFlowDefinitionsResponse_nextToken - A token to resume pagination.

$sel:httpStatus:ListFlowDefinitionsResponse', listFlowDefinitionsResponse_httpStatus - The response's http status code.

$sel:flowDefinitionSummaries:ListFlowDefinitionsResponse', listFlowDefinitionsResponse_flowDefinitionSummaries - An array of objects describing the flow definitions.

Response Lenses

listFlowDefinitionsResponse_nextToken :: Lens' ListFlowDefinitionsResponse (Maybe Text)

A token to resume pagination.

listFlowDefinitionsResponse_httpStatus :: Lens' ListFlowDefinitionsResponse Int

The response's http status code.

listFlowDefinitionsResponse_flowDefinitionSummaries :: Lens' ListFlowDefinitionsResponse [FlowDefinitionSummary]

An array of objects describing the flow definitions.