Network.AWS.SageMaker.ListHyperParameterTuningJobs

Description

Gets a list of HyperParameterTuningJobSummary objects that describe the hyperparameter tuning jobs launched in your account.

This operation returns paginated results.

Creating a Request

data ListHyperParameterTuningJobs

See: newListHyperParameterTuningJobs smart constructor.

Constructors

ListHyperParameterTuningJobs'

Fields

lastModifiedTimeBefore :: Maybe POSIX

A filter that returns only tuning jobs that were modified before the specified time.

nextToken :: Maybe Text

If the result of the previous ListHyperParameterTuningJobs request was truncated, the response includes a NextToken. To retrieve the next set of tuning jobs, use the token in the next request.

sortOrder :: Maybe SortOrder

The sort order for results. The default is Ascending.

nameContains :: Maybe Text

A string in the tuning job name. This filter returns only tuning jobs whose name contains the specified string.

maxResults :: Maybe Natural

The maximum number of tuning jobs to return. The default value is 10.

creationTimeBefore :: Maybe POSIX

A filter that returns only tuning jobs that were created before the specified time.

lastModifiedTimeAfter :: Maybe POSIX

A filter that returns only tuning jobs that were modified after the specified time.

sortBy :: Maybe HyperParameterTuningJobSortByOptions

The field to sort results by. The default is Name.

statusEquals :: Maybe HyperParameterTuningJobStatus

A filter that returns only tuning jobs with the specified status.

creationTimeAfter :: Maybe POSIX

A filter that returns only tuning jobs that were created after the specified time.

newListHyperParameterTuningJobs :: ListHyperParameterTuningJobs

Create a value of ListHyperParameterTuningJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedTimeBefore:ListHyperParameterTuningJobs', listHyperParameterTuningJobs_lastModifiedTimeBefore - A filter that returns only tuning jobs that were modified before the specified time.

$sel:nextToken:ListHyperParameterTuningJobs', listHyperParameterTuningJobs_nextToken - If the result of the previous ListHyperParameterTuningJobs request was truncated, the response includes a NextToken. To retrieve the next set of tuning jobs, use the token in the next request.

$sel:sortOrder:ListHyperParameterTuningJobs', listHyperParameterTuningJobs_sortOrder - The sort order for results. The default is Ascending.

$sel:nameContains:ListHyperParameterTuningJobs', listHyperParameterTuningJobs_nameContains - A string in the tuning job name. This filter returns only tuning jobs whose name contains the specified string.

$sel:maxResults:ListHyperParameterTuningJobs', listHyperParameterTuningJobs_maxResults - The maximum number of tuning jobs to return. The default value is 10.

$sel:creationTimeBefore:ListHyperParameterTuningJobs', listHyperParameterTuningJobs_creationTimeBefore - A filter that returns only tuning jobs that were created before the specified time.

$sel:lastModifiedTimeAfter:ListHyperParameterTuningJobs', listHyperParameterTuningJobs_lastModifiedTimeAfter - A filter that returns only tuning jobs that were modified after the specified time.

$sel:sortBy:ListHyperParameterTuningJobs', listHyperParameterTuningJobs_sortBy - The field to sort results by. The default is Name.

$sel:statusEquals:ListHyperParameterTuningJobs', listHyperParameterTuningJobs_statusEquals - A filter that returns only tuning jobs with the specified status.

$sel:creationTimeAfter:ListHyperParameterTuningJobs', listHyperParameterTuningJobs_creationTimeAfter - A filter that returns only tuning jobs that were created after the specified time.

Request Lenses

listHyperParameterTuningJobs_lastModifiedTimeBefore :: Lens' ListHyperParameterTuningJobs (Maybe UTCTime)

A filter that returns only tuning jobs that were modified before the specified time.

listHyperParameterTuningJobs_nextToken :: Lens' ListHyperParameterTuningJobs (Maybe Text)

If the result of the previous ListHyperParameterTuningJobs request was truncated, the response includes a NextToken. To retrieve the next set of tuning jobs, use the token in the next request.

listHyperParameterTuningJobs_sortOrder :: Lens' ListHyperParameterTuningJobs (Maybe SortOrder)

The sort order for results. The default is Ascending.

listHyperParameterTuningJobs_nameContains :: Lens' ListHyperParameterTuningJobs (Maybe Text)

A string in the tuning job name. This filter returns only tuning jobs whose name contains the specified string.

listHyperParameterTuningJobs_maxResults :: Lens' ListHyperParameterTuningJobs (Maybe Natural)

The maximum number of tuning jobs to return. The default value is 10.

listHyperParameterTuningJobs_creationTimeBefore :: Lens' ListHyperParameterTuningJobs (Maybe UTCTime)

A filter that returns only tuning jobs that were created before the specified time.

listHyperParameterTuningJobs_lastModifiedTimeAfter :: Lens' ListHyperParameterTuningJobs (Maybe UTCTime)

A filter that returns only tuning jobs that were modified after the specified time.

listHyperParameterTuningJobs_sortBy :: Lens' ListHyperParameterTuningJobs (Maybe HyperParameterTuningJobSortByOptions)

The field to sort results by. The default is Name.

listHyperParameterTuningJobs_statusEquals :: Lens' ListHyperParameterTuningJobs (Maybe HyperParameterTuningJobStatus)

A filter that returns only tuning jobs with the specified status.

listHyperParameterTuningJobs_creationTimeAfter :: Lens' ListHyperParameterTuningJobs (Maybe UTCTime)

A filter that returns only tuning jobs that were created after the specified time.

Destructuring the Response

data ListHyperParameterTuningJobsResponse

See: newListHyperParameterTuningJobsResponse smart constructor.

Constructors

ListHyperParameterTuningJobsResponse'

Fields

nextToken :: Maybe Text

If the result of this ListHyperParameterTuningJobs request was truncated, the response includes a NextToken. To retrieve the next set of tuning jobs, use the token in the next request.

httpStatus :: Int

The response's http status code.

hyperParameterTuningJobSummaries :: [HyperParameterTuningJobSummary]

A list of HyperParameterTuningJobSummary objects that describe the tuning jobs that the ListHyperParameterTuningJobs request returned.

newListHyperParameterTuningJobsResponse

Arguments

:: Int
$sel:httpStatus:ListHyperParameterTuningJobsResponse'

-> ListHyperParameterTuningJobsResponse

Create a value of ListHyperParameterTuningJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListHyperParameterTuningJobs', listHyperParameterTuningJobsResponse_nextToken - If the result of this ListHyperParameterTuningJobs request was truncated, the response includes a NextToken. To retrieve the next set of tuning jobs, use the token in the next request.

$sel:httpStatus:ListHyperParameterTuningJobsResponse', listHyperParameterTuningJobsResponse_httpStatus - The response's http status code.

$sel:hyperParameterTuningJobSummaries:ListHyperParameterTuningJobsResponse', listHyperParameterTuningJobsResponse_hyperParameterTuningJobSummaries - A list of HyperParameterTuningJobSummary objects that describe the tuning jobs that the ListHyperParameterTuningJobs request returned.

Response Lenses

listHyperParameterTuningJobsResponse_nextToken :: Lens' ListHyperParameterTuningJobsResponse (Maybe Text)

If the result of this ListHyperParameterTuningJobs request was truncated, the response includes a NextToken. To retrieve the next set of tuning jobs, use the token in the next request.

listHyperParameterTuningJobsResponse_httpStatus :: Lens' ListHyperParameterTuningJobsResponse Int

The response's http status code.

listHyperParameterTuningJobsResponse_hyperParameterTuningJobSummaries :: Lens' ListHyperParameterTuningJobsResponse [HyperParameterTuningJobSummary]

A list of HyperParameterTuningJobSummary objects that describe the tuning jobs that the ListHyperParameterTuningJobs request returned.