Network.AWS.SageMaker.ListLabelingJobs

Description

Gets a list of labeling jobs.

This operation returns paginated results.

Creating a Request

data ListLabelingJobs

See: newListLabelingJobs smart constructor.

Constructors

ListLabelingJobs'

Fields

nameContains :: Maybe Text

A string in the labeling job name. This filter returns only labeling jobs whose name contains the specified string.

lastModifiedTimeBefore :: Maybe POSIX

A filter that returns only labeling jobs modified before the specified time (timestamp).

creationTimeAfter :: Maybe POSIX

A filter that returns only labeling jobs created after the specified time (timestamp).

nextToken :: Maybe Text

If the result of the previous ListLabelingJobs request was truncated, the response includes a NextToken. To retrieve the next set of labeling jobs, use the token in the next request.

sortOrder :: Maybe SortOrder

The sort order for results. The default is Ascending.

lastModifiedTimeAfter :: Maybe POSIX

A filter that returns only labeling jobs modified after the specified time (timestamp).

creationTimeBefore :: Maybe POSIX

A filter that returns only labeling jobs created before the specified time (timestamp).

statusEquals :: Maybe LabelingJobStatus

A filter that retrieves only labeling jobs with a specific status.

maxResults :: Maybe Natural

The maximum number of labeling jobs to return in each page of the response.

sortBy :: Maybe SortBy

The field to sort results by. The default is CreationTime.

newListLabelingJobs :: ListLabelingJobs

Create a value of ListLabelingJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nameContains:ListLabelingJobs', listLabelingJobs_nameContains - A string in the labeling job name. This filter returns only labeling jobs whose name contains the specified string.

$sel:lastModifiedTimeBefore:ListLabelingJobs', listLabelingJobs_lastModifiedTimeBefore - A filter that returns only labeling jobs modified before the specified time (timestamp).

$sel:creationTimeAfter:ListLabelingJobs', listLabelingJobs_creationTimeAfter - A filter that returns only labeling jobs created after the specified time (timestamp).

$sel:nextToken:ListLabelingJobs', listLabelingJobs_nextToken - If the result of the previous ListLabelingJobs request was truncated, the response includes a NextToken. To retrieve the next set of labeling jobs, use the token in the next request.

$sel:sortOrder:ListLabelingJobs', listLabelingJobs_sortOrder - The sort order for results. The default is Ascending.

$sel:lastModifiedTimeAfter:ListLabelingJobs', listLabelingJobs_lastModifiedTimeAfter - A filter that returns only labeling jobs modified after the specified time (timestamp).

$sel:creationTimeBefore:ListLabelingJobs', listLabelingJobs_creationTimeBefore - A filter that returns only labeling jobs created before the specified time (timestamp).

$sel:statusEquals:ListLabelingJobs', listLabelingJobs_statusEquals - A filter that retrieves only labeling jobs with a specific status.

$sel:maxResults:ListLabelingJobs', listLabelingJobs_maxResults - The maximum number of labeling jobs to return in each page of the response.

$sel:sortBy:ListLabelingJobs', listLabelingJobs_sortBy - The field to sort results by. The default is CreationTime.

Request Lenses

listLabelingJobs_nameContains :: Lens' ListLabelingJobs (Maybe Text)

A string in the labeling job name. This filter returns only labeling jobs whose name contains the specified string.

listLabelingJobs_lastModifiedTimeBefore :: Lens' ListLabelingJobs (Maybe UTCTime)

A filter that returns only labeling jobs modified before the specified time (timestamp).

listLabelingJobs_creationTimeAfter :: Lens' ListLabelingJobs (Maybe UTCTime)

A filter that returns only labeling jobs created after the specified time (timestamp).

listLabelingJobs_nextToken :: Lens' ListLabelingJobs (Maybe Text)

If the result of the previous ListLabelingJobs request was truncated, the response includes a NextToken. To retrieve the next set of labeling jobs, use the token in the next request.

listLabelingJobs_sortOrder :: Lens' ListLabelingJobs (Maybe SortOrder)

The sort order for results. The default is Ascending.

listLabelingJobs_lastModifiedTimeAfter :: Lens' ListLabelingJobs (Maybe UTCTime)

A filter that returns only labeling jobs modified after the specified time (timestamp).

listLabelingJobs_creationTimeBefore :: Lens' ListLabelingJobs (Maybe UTCTime)

A filter that returns only labeling jobs created before the specified time (timestamp).

listLabelingJobs_statusEquals :: Lens' ListLabelingJobs (Maybe LabelingJobStatus)

A filter that retrieves only labeling jobs with a specific status.

listLabelingJobs_maxResults :: Lens' ListLabelingJobs (Maybe Natural)

The maximum number of labeling jobs to return in each page of the response.

listLabelingJobs_sortBy :: Lens' ListLabelingJobs (Maybe SortBy)

The field to sort results by. The default is CreationTime.

Destructuring the Response

data ListLabelingJobsResponse

See: newListLabelingJobsResponse smart constructor.

Constructors

ListLabelingJobsResponse'

Fields

labelingJobSummaryList :: Maybe [LabelingJobSummary]

An array of LabelingJobSummary objects, each describing a labeling job.

nextToken :: Maybe Text

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of labeling jobs, use it in the subsequent request.

httpStatus :: Int

The response's http status code.

newListLabelingJobsResponse

Arguments

:: Int
$sel:httpStatus:ListLabelingJobsResponse'

-> ListLabelingJobsResponse

Create a value of ListLabelingJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:labelingJobSummaryList:ListLabelingJobsResponse', listLabelingJobsResponse_labelingJobSummaryList - An array of LabelingJobSummary objects, each describing a labeling job.

$sel:nextToken:ListLabelingJobs', listLabelingJobsResponse_nextToken - If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of labeling jobs, use it in the subsequent request.

$sel:httpStatus:ListLabelingJobsResponse', listLabelingJobsResponse_httpStatus - The response's http status code.

Response Lenses

listLabelingJobsResponse_labelingJobSummaryList :: Lens' ListLabelingJobsResponse (Maybe [LabelingJobSummary])

An array of LabelingJobSummary objects, each describing a labeling job.

listLabelingJobsResponse_nextToken :: Lens' ListLabelingJobsResponse (Maybe Text)

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of labeling jobs, use it in the subsequent request.

listLabelingJobsResponse_httpStatus :: Lens' ListLabelingJobsResponse Int

The response's http status code.