Network.AWS.SageMaker.ListLabelingJobsForWorkteam

Description

Gets a list of labeling jobs assigned to a specified work team.

This operation returns paginated results.

Creating a Request

data ListLabelingJobsForWorkteam

See: newListLabelingJobsForWorkteam smart constructor.

Constructors

ListLabelingJobsForWorkteam'

Fields

jobReferenceCodeContains :: Maybe Text

A filter the limits jobs to only the ones whose job reference code contains the specified string.

creationTimeAfter :: Maybe POSIX

A filter that returns only labeling jobs created after the specified time (timestamp).

nextToken :: Maybe Text

If the result of the previous ListLabelingJobsForWorkteam request was truncated, the response includes a NextToken. To retrieve the next set of labeling jobs, use the token in the next request.

sortOrder :: Maybe SortOrder

The sort order for results. The default is Ascending.

creationTimeBefore :: Maybe POSIX

A filter that returns only labeling jobs created before the specified time (timestamp).

maxResults :: Maybe Natural

The maximum number of labeling jobs to return in each page of the response.

sortBy :: Maybe ListLabelingJobsForWorkteamSortByOptions

The field to sort results by. The default is CreationTime.

workteamArn :: Text

The Amazon Resource Name (ARN) of the work team for which you want to see labeling jobs for.

newListLabelingJobsForWorkteam

Arguments

:: Text
$sel:workteamArn:ListLabelingJobsForWorkteam'

-> ListLabelingJobsForWorkteam

Create a value of ListLabelingJobsForWorkteam with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobReferenceCodeContains:ListLabelingJobsForWorkteam', listLabelingJobsForWorkteam_jobReferenceCodeContains - A filter the limits jobs to only the ones whose job reference code contains the specified string.

$sel:creationTimeAfter:ListLabelingJobsForWorkteam', listLabelingJobsForWorkteam_creationTimeAfter - A filter that returns only labeling jobs created after the specified time (timestamp).

$sel:nextToken:ListLabelingJobsForWorkteam', listLabelingJobsForWorkteam_nextToken - If the result of the previous ListLabelingJobsForWorkteam request was truncated, the response includes a NextToken. To retrieve the next set of labeling jobs, use the token in the next request.

$sel:sortOrder:ListLabelingJobsForWorkteam', listLabelingJobsForWorkteam_sortOrder - The sort order for results. The default is Ascending.

$sel:creationTimeBefore:ListLabelingJobsForWorkteam', listLabelingJobsForWorkteam_creationTimeBefore - A filter that returns only labeling jobs created before the specified time (timestamp).

$sel:maxResults:ListLabelingJobsForWorkteam', listLabelingJobsForWorkteam_maxResults - The maximum number of labeling jobs to return in each page of the response.

$sel:sortBy:ListLabelingJobsForWorkteam', listLabelingJobsForWorkteam_sortBy - The field to sort results by. The default is CreationTime.

$sel:workteamArn:ListLabelingJobsForWorkteam', listLabelingJobsForWorkteam_workteamArn - The Amazon Resource Name (ARN) of the work team for which you want to see labeling jobs for.

Request Lenses

listLabelingJobsForWorkteam_jobReferenceCodeContains :: Lens' ListLabelingJobsForWorkteam (Maybe Text)

A filter the limits jobs to only the ones whose job reference code contains the specified string.

listLabelingJobsForWorkteam_creationTimeAfter :: Lens' ListLabelingJobsForWorkteam (Maybe UTCTime)

A filter that returns only labeling jobs created after the specified time (timestamp).

listLabelingJobsForWorkteam_nextToken :: Lens' ListLabelingJobsForWorkteam (Maybe Text)

If the result of the previous ListLabelingJobsForWorkteam request was truncated, the response includes a NextToken. To retrieve the next set of labeling jobs, use the token in the next request.

listLabelingJobsForWorkteam_sortOrder :: Lens' ListLabelingJobsForWorkteam (Maybe SortOrder)

The sort order for results. The default is Ascending.

listLabelingJobsForWorkteam_creationTimeBefore :: Lens' ListLabelingJobsForWorkteam (Maybe UTCTime)

A filter that returns only labeling jobs created before the specified time (timestamp).

listLabelingJobsForWorkteam_maxResults :: Lens' ListLabelingJobsForWorkteam (Maybe Natural)

The maximum number of labeling jobs to return in each page of the response.

listLabelingJobsForWorkteam_sortBy :: Lens' ListLabelingJobsForWorkteam (Maybe ListLabelingJobsForWorkteamSortByOptions)

The field to sort results by. The default is CreationTime.

listLabelingJobsForWorkteam_workteamArn :: Lens' ListLabelingJobsForWorkteam Text

The Amazon Resource Name (ARN) of the work team for which you want to see labeling jobs for.

Destructuring the Response

data ListLabelingJobsForWorkteamResponse

See: newListLabelingJobsForWorkteamResponse smart constructor.

Constructors

ListLabelingJobsForWorkteamResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of labeling jobs, use it in the subsequent request.

httpStatus :: Int

The response's http status code.

labelingJobSummaryList :: [LabelingJobForWorkteamSummary]

An array of LabelingJobSummary objects, each describing a labeling job.

newListLabelingJobsForWorkteamResponse

Arguments

:: Int
$sel:httpStatus:ListLabelingJobsForWorkteamResponse'

-> ListLabelingJobsForWorkteamResponse

Create a value of ListLabelingJobsForWorkteamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListLabelingJobsForWorkteam', listLabelingJobsForWorkteamResponse_nextToken - If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of labeling jobs, use it in the subsequent request.

$sel:httpStatus:ListLabelingJobsForWorkteamResponse', listLabelingJobsForWorkteamResponse_httpStatus - The response's http status code.

$sel:labelingJobSummaryList:ListLabelingJobsForWorkteamResponse', listLabelingJobsForWorkteamResponse_labelingJobSummaryList - An array of LabelingJobSummary objects, each describing a labeling job.

Response Lenses

listLabelingJobsForWorkteamResponse_nextToken :: Lens' ListLabelingJobsForWorkteamResponse (Maybe Text)

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of labeling jobs, use it in the subsequent request.

listLabelingJobsForWorkteamResponse_httpStatus :: Lens' ListLabelingJobsForWorkteamResponse Int

The response's http status code.

listLabelingJobsForWorkteamResponse_labelingJobSummaryList :: Lens' ListLabelingJobsForWorkteamResponse [LabelingJobForWorkteamSummary]

An array of LabelingJobSummary objects, each describing a labeling job.