Network.AWS.SageMaker.ListModelExplainabilityJobDefinitions

Description

Lists model explainability job definitions that satisfy various filters.

This operation returns paginated results.

Creating a Request

data ListModelExplainabilityJobDefinitions

See: newListModelExplainabilityJobDefinitions smart constructor.

Constructors

ListModelExplainabilityJobDefinitions'

Fields

nameContains :: Maybe Text

Filter for model explainability jobs whose name contains a specified string.

endpointName :: Maybe Text

Name of the endpoint to monitor for model explainability.

creationTimeAfter :: Maybe POSIX

A filter that returns only model explainability jobs created after a specified time.

nextToken :: Maybe Text

The token returned if the response is truncated. To retrieve the next set of job executions, use it in the next request.

sortOrder :: Maybe SortOrder

Whether to sort the results in Ascending or Descending order. The default is Descending.

creationTimeBefore :: Maybe POSIX

A filter that returns only model explainability jobs created before a specified time.

maxResults :: Maybe Natural

The maximum number of jobs to return in the response. The default value is 10.

sortBy :: Maybe MonitoringJobDefinitionSortKey

Whether to sort results by the Name or CreationTime field. The default is CreationTime.

newListModelExplainabilityJobDefinitions :: ListModelExplainabilityJobDefinitions

Create a value of ListModelExplainabilityJobDefinitions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nameContains:ListModelExplainabilityJobDefinitions', listModelExplainabilityJobDefinitions_nameContains - Filter for model explainability jobs whose name contains a specified string.

$sel:endpointName:ListModelExplainabilityJobDefinitions', listModelExplainabilityJobDefinitions_endpointName - Name of the endpoint to monitor for model explainability.

$sel:creationTimeAfter:ListModelExplainabilityJobDefinitions', listModelExplainabilityJobDefinitions_creationTimeAfter - A filter that returns only model explainability jobs created after a specified time.

$sel:nextToken:ListModelExplainabilityJobDefinitions', listModelExplainabilityJobDefinitions_nextToken - The token returned if the response is truncated. To retrieve the next set of job executions, use it in the next request.

$sel:sortOrder:ListModelExplainabilityJobDefinitions', listModelExplainabilityJobDefinitions_sortOrder - Whether to sort the results in Ascending or Descending order. The default is Descending.

$sel:creationTimeBefore:ListModelExplainabilityJobDefinitions', listModelExplainabilityJobDefinitions_creationTimeBefore - A filter that returns only model explainability jobs created before a specified time.

$sel:maxResults:ListModelExplainabilityJobDefinitions', listModelExplainabilityJobDefinitions_maxResults - The maximum number of jobs to return in the response. The default value is 10.

$sel:sortBy:ListModelExplainabilityJobDefinitions', listModelExplainabilityJobDefinitions_sortBy - Whether to sort results by the Name or CreationTime field. The default is CreationTime.

Request Lenses

listModelExplainabilityJobDefinitions_nameContains :: Lens' ListModelExplainabilityJobDefinitions (Maybe Text)

Filter for model explainability jobs whose name contains a specified string.

listModelExplainabilityJobDefinitions_endpointName :: Lens' ListModelExplainabilityJobDefinitions (Maybe Text)

Name of the endpoint to monitor for model explainability.

listModelExplainabilityJobDefinitions_creationTimeAfter :: Lens' ListModelExplainabilityJobDefinitions (Maybe UTCTime)

A filter that returns only model explainability jobs created after a specified time.

listModelExplainabilityJobDefinitions_nextToken :: Lens' ListModelExplainabilityJobDefinitions (Maybe Text)

The token returned if the response is truncated. To retrieve the next set of job executions, use it in the next request.

listModelExplainabilityJobDefinitions_sortOrder :: Lens' ListModelExplainabilityJobDefinitions (Maybe SortOrder)

Whether to sort the results in Ascending or Descending order. The default is Descending.

listModelExplainabilityJobDefinitions_creationTimeBefore :: Lens' ListModelExplainabilityJobDefinitions (Maybe UTCTime)

A filter that returns only model explainability jobs created before a specified time.

listModelExplainabilityJobDefinitions_maxResults :: Lens' ListModelExplainabilityJobDefinitions (Maybe Natural)

The maximum number of jobs to return in the response. The default value is 10.

listModelExplainabilityJobDefinitions_sortBy :: Lens' ListModelExplainabilityJobDefinitions (Maybe MonitoringJobDefinitionSortKey)

Whether to sort results by the Name or CreationTime field. The default is CreationTime.

Destructuring the Response

data ListModelExplainabilityJobDefinitionsResponse

See: newListModelExplainabilityJobDefinitionsResponse smart constructor.

Constructors

ListModelExplainabilityJobDefinitionsResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of jobs, use it in the subsequent request.

httpStatus :: Int

The response's http status code.

jobDefinitionSummaries :: [MonitoringJobDefinitionSummary]

A JSON array in which each element is a summary for a explainability bias jobs.

newListModelExplainabilityJobDefinitionsResponse

Arguments

:: Int
$sel:httpStatus:ListModelExplainabilityJobDefinitionsResponse'

-> ListModelExplainabilityJobDefinitionsResponse

Create a value of ListModelExplainabilityJobDefinitionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListModelExplainabilityJobDefinitions', listModelExplainabilityJobDefinitionsResponse_nextToken - If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of jobs, use it in the subsequent request.

$sel:httpStatus:ListModelExplainabilityJobDefinitionsResponse', listModelExplainabilityJobDefinitionsResponse_httpStatus - The response's http status code.

$sel:jobDefinitionSummaries:ListModelExplainabilityJobDefinitionsResponse', listModelExplainabilityJobDefinitionsResponse_jobDefinitionSummaries - A JSON array in which each element is a summary for a explainability bias jobs.

Response Lenses

listModelExplainabilityJobDefinitionsResponse_nextToken :: Lens' ListModelExplainabilityJobDefinitionsResponse (Maybe Text)

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of jobs, use it in the subsequent request.

listModelExplainabilityJobDefinitionsResponse_httpStatus :: Lens' ListModelExplainabilityJobDefinitionsResponse Int

The response's http status code.

listModelExplainabilityJobDefinitionsResponse_jobDefinitionSummaries :: Lens' ListModelExplainabilityJobDefinitionsResponse [MonitoringJobDefinitionSummary]

A JSON array in which each element is a summary for a explainability bias jobs.