Network.AWS.SageMaker.ListModelPackages

Description

Lists the model packages that have been created.

This operation returns paginated results.

Creating a Request

data ListModelPackages

See: newListModelPackages smart constructor.

Constructors

ListModelPackages'

Fields

nameContains :: Maybe Text

A string in the model package name. This filter returns only model packages whose name contains the specified string.

modelApprovalStatus :: Maybe ModelApprovalStatus

A filter that returns only the model packages with the specified approval status.

modelPackageType :: Maybe ModelPackageType

A filter that returns onlyl the model packages of the specified type. This can be one of the following values.

VERSIONED - List only versioned models.

UNVERSIONED - List only unversioined models.

BOTH - List both versioned and unversioned models.

creationTimeAfter :: Maybe POSIX

A filter that returns only model packages created after the specified time (timestamp).

nextToken :: Maybe Text

If the response to a previous ListModelPackages request was truncated, the response includes a NextToken. To retrieve the next set of model packages, use the token in the next request.

sortOrder :: Maybe SortOrder

The sort order for the results. The default is Ascending.

modelPackageGroupName :: Maybe Text

A filter that returns only model versions that belong to the specified model group.

creationTimeBefore :: Maybe POSIX

A filter that returns only model packages created before the specified time (timestamp).

maxResults :: Maybe Natural

The maximum number of model packages to return in the response.

sortBy :: Maybe ModelPackageSortBy

The parameter by which to sort the results. The default is CreationTime.

newListModelPackages :: ListModelPackages

Create a value of ListModelPackages with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nameContains:ListModelPackages', listModelPackages_nameContains - A string in the model package name. This filter returns only model packages whose name contains the specified string.

$sel:modelApprovalStatus:ListModelPackages', listModelPackages_modelApprovalStatus - A filter that returns only the model packages with the specified approval status.

$sel:modelPackageType:ListModelPackages', listModelPackages_modelPackageType - A filter that returns onlyl the model packages of the specified type. This can be one of the following values.

VERSIONED - List only versioned models.

UNVERSIONED - List only unversioined models.

BOTH - List both versioned and unversioned models.

$sel:creationTimeAfter:ListModelPackages', listModelPackages_creationTimeAfter - A filter that returns only model packages created after the specified time (timestamp).

$sel:nextToken:ListModelPackages', listModelPackages_nextToken - If the response to a previous ListModelPackages request was truncated, the response includes a NextToken. To retrieve the next set of model packages, use the token in the next request.

$sel:sortOrder:ListModelPackages', listModelPackages_sortOrder - The sort order for the results. The default is Ascending.

$sel:modelPackageGroupName:ListModelPackages', listModelPackages_modelPackageGroupName - A filter that returns only model versions that belong to the specified model group.

$sel:creationTimeBefore:ListModelPackages', listModelPackages_creationTimeBefore - A filter that returns only model packages created before the specified time (timestamp).

$sel:maxResults:ListModelPackages', listModelPackages_maxResults - The maximum number of model packages to return in the response.

$sel:sortBy:ListModelPackages', listModelPackages_sortBy - The parameter by which to sort the results. The default is CreationTime.

Request Lenses

listModelPackages_nameContains :: Lens' ListModelPackages (Maybe Text)

A string in the model package name. This filter returns only model packages whose name contains the specified string.

listModelPackages_modelApprovalStatus :: Lens' ListModelPackages (Maybe ModelApprovalStatus)

A filter that returns only the model packages with the specified approval status.

listModelPackages_modelPackageType :: Lens' ListModelPackages (Maybe ModelPackageType)

A filter that returns onlyl the model packages of the specified type. This can be one of the following values.

VERSIONED - List only versioned models.

UNVERSIONED - List only unversioined models.

BOTH - List both versioned and unversioned models.

listModelPackages_creationTimeAfter :: Lens' ListModelPackages (Maybe UTCTime)

A filter that returns only model packages created after the specified time (timestamp).

listModelPackages_nextToken :: Lens' ListModelPackages (Maybe Text)

If the response to a previous ListModelPackages request was truncated, the response includes a NextToken. To retrieve the next set of model packages, use the token in the next request.

listModelPackages_sortOrder :: Lens' ListModelPackages (Maybe SortOrder)

The sort order for the results. The default is Ascending.

listModelPackages_modelPackageGroupName :: Lens' ListModelPackages (Maybe Text)

A filter that returns only model versions that belong to the specified model group.

listModelPackages_creationTimeBefore :: Lens' ListModelPackages (Maybe UTCTime)

A filter that returns only model packages created before the specified time (timestamp).

listModelPackages_maxResults :: Lens' ListModelPackages (Maybe Natural)

The maximum number of model packages to return in the response.

listModelPackages_sortBy :: Lens' ListModelPackages (Maybe ModelPackageSortBy)

The parameter by which to sort the results. The default is CreationTime.

Destructuring the Response

data ListModelPackagesResponse

See: newListModelPackagesResponse smart constructor.

Constructors

ListModelPackagesResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of model packages, use it in the subsequent request.

httpStatus :: Int

The response's http status code.

modelPackageSummaryList :: [ModelPackageSummary]

An array of ModelPackageSummary objects, each of which lists a model package.

newListModelPackagesResponse

Arguments

:: Int
$sel:httpStatus:ListModelPackagesResponse'

-> ListModelPackagesResponse

Create a value of ListModelPackagesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListModelPackages', listModelPackagesResponse_nextToken - If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of model packages, use it in the subsequent request.

$sel:httpStatus:ListModelPackagesResponse', listModelPackagesResponse_httpStatus - The response's http status code.

$sel:modelPackageSummaryList:ListModelPackagesResponse', listModelPackagesResponse_modelPackageSummaryList - An array of ModelPackageSummary objects, each of which lists a model package.

Response Lenses

listModelPackagesResponse_nextToken :: Lens' ListModelPackagesResponse (Maybe Text)

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of model packages, use it in the subsequent request.

listModelPackagesResponse_httpStatus :: Lens' ListModelPackagesResponse Int

The response's http status code.

listModelPackagesResponse_modelPackageSummaryList :: Lens' ListModelPackagesResponse [ModelPackageSummary]

An array of ModelPackageSummary objects, each of which lists a model package.