Network.AWS.SageMaker.ListModelQualityJobDefinitions

Description

Gets a list of model quality monitoring job definitions in your account.

This operation returns paginated results.

Creating a Request

data ListModelQualityJobDefinitions

See: newListModelQualityJobDefinitions smart constructor.

Constructors

ListModelQualityJobDefinitions'

Fields

nextToken :: Maybe Text

If the result of the previous ListModelQualityJobDefinitions request was truncated, the response includes a NextToken. To retrieve the next set of model quality monitoring job definitions, use the token in the next request.

sortOrder :: Maybe SortOrder

The sort order for results. The default is Descending.

endpointName :: Maybe Text

A filter that returns only model quality monitoring job definitions that are associated with the specified endpoint.

nameContains :: Maybe Text

A string in the transform job name. This filter returns only model quality monitoring job definitions whose name contains the specified string.

maxResults :: Maybe Natural

The maximum number of results to return in a call to ListModelQualityJobDefinitions.

creationTimeBefore :: Maybe POSIX

A filter that returns only model quality monitoring job definitions created before the specified time.

sortBy :: Maybe MonitoringJobDefinitionSortKey

The field to sort results by. The default is CreationTime.

creationTimeAfter :: Maybe POSIX

A filter that returns only model quality monitoring job definitions created after the specified time.

newListModelQualityJobDefinitions :: ListModelQualityJobDefinitions

Create a value of ListModelQualityJobDefinitions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListModelQualityJobDefinitions', listModelQualityJobDefinitions_nextToken - If the result of the previous ListModelQualityJobDefinitions request was truncated, the response includes a NextToken. To retrieve the next set of model quality monitoring job definitions, use the token in the next request.

$sel:sortOrder:ListModelQualityJobDefinitions', listModelQualityJobDefinitions_sortOrder - The sort order for results. The default is Descending.

$sel:endpointName:ListModelQualityJobDefinitions', listModelQualityJobDefinitions_endpointName - A filter that returns only model quality monitoring job definitions that are associated with the specified endpoint.

$sel:nameContains:ListModelQualityJobDefinitions', listModelQualityJobDefinitions_nameContains - A string in the transform job name. This filter returns only model quality monitoring job definitions whose name contains the specified string.

$sel:maxResults:ListModelQualityJobDefinitions', listModelQualityJobDefinitions_maxResults - The maximum number of results to return in a call to ListModelQualityJobDefinitions.

$sel:creationTimeBefore:ListModelQualityJobDefinitions', listModelQualityJobDefinitions_creationTimeBefore - A filter that returns only model quality monitoring job definitions created before the specified time.

$sel:sortBy:ListModelQualityJobDefinitions', listModelQualityJobDefinitions_sortBy - The field to sort results by. The default is CreationTime.

$sel:creationTimeAfter:ListModelQualityJobDefinitions', listModelQualityJobDefinitions_creationTimeAfter - A filter that returns only model quality monitoring job definitions created after the specified time.

Request Lenses

listModelQualityJobDefinitions_nextToken :: Lens' ListModelQualityJobDefinitions (Maybe Text)

If the result of the previous ListModelQualityJobDefinitions request was truncated, the response includes a NextToken. To retrieve the next set of model quality monitoring job definitions, use the token in the next request.

listModelQualityJobDefinitions_sortOrder :: Lens' ListModelQualityJobDefinitions (Maybe SortOrder)

The sort order for results. The default is Descending.

listModelQualityJobDefinitions_endpointName :: Lens' ListModelQualityJobDefinitions (Maybe Text)

A filter that returns only model quality monitoring job definitions that are associated with the specified endpoint.

listModelQualityJobDefinitions_nameContains :: Lens' ListModelQualityJobDefinitions (Maybe Text)

A string in the transform job name. This filter returns only model quality monitoring job definitions whose name contains the specified string.

listModelQualityJobDefinitions_maxResults :: Lens' ListModelQualityJobDefinitions (Maybe Natural)

The maximum number of results to return in a call to ListModelQualityJobDefinitions.

listModelQualityJobDefinitions_creationTimeBefore :: Lens' ListModelQualityJobDefinitions (Maybe UTCTime)

A filter that returns only model quality monitoring job definitions created before the specified time.

listModelQualityJobDefinitions_sortBy :: Lens' ListModelQualityJobDefinitions (Maybe MonitoringJobDefinitionSortKey)

The field to sort results by. The default is CreationTime.

listModelQualityJobDefinitions_creationTimeAfter :: Lens' ListModelQualityJobDefinitions (Maybe UTCTime)

A filter that returns only model quality monitoring job definitions created after the specified time.

Destructuring the Response

data ListModelQualityJobDefinitionsResponse

See: newListModelQualityJobDefinitionsResponse smart constructor.

Constructors

ListModelQualityJobDefinitionsResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of model quality monitoring job definitions, use it in the next request.

httpStatus :: Int

The response's http status code.

jobDefinitionSummaries :: [MonitoringJobDefinitionSummary]

A list of summaries of model quality monitoring job definitions.

newListModelQualityJobDefinitionsResponse

Arguments

:: Int
$sel:httpStatus:ListModelQualityJobDefinitionsResponse'

-> ListModelQualityJobDefinitionsResponse

Create a value of ListModelQualityJobDefinitionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListModelQualityJobDefinitions', listModelQualityJobDefinitionsResponse_nextToken - If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of model quality monitoring job definitions, use it in the next request.

$sel:httpStatus:ListModelQualityJobDefinitionsResponse', listModelQualityJobDefinitionsResponse_httpStatus - The response's http status code.

$sel:jobDefinitionSummaries:ListModelQualityJobDefinitionsResponse', listModelQualityJobDefinitionsResponse_jobDefinitionSummaries - A list of summaries of model quality monitoring job definitions.

Response Lenses

listModelQualityJobDefinitionsResponse_nextToken :: Lens' ListModelQualityJobDefinitionsResponse (Maybe Text)

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of model quality monitoring job definitions, use it in the next request.

listModelQualityJobDefinitionsResponse_httpStatus :: Lens' ListModelQualityJobDefinitionsResponse Int

The response's http status code.

listModelQualityJobDefinitionsResponse_jobDefinitionSummaries :: Lens' ListModelQualityJobDefinitionsResponse [MonitoringJobDefinitionSummary]

A list of summaries of model quality monitoring job definitions.