Network.AWS.SageMaker.ListModels

Description

Lists models created with the CreateModel API.

This operation returns paginated results.

Creating a Request

data ListModels

See: newListModels smart constructor.

Constructors

ListModels'

Fields

nextToken :: Maybe Text

If the response to a previous ListModels request was truncated, the response includes a NextToken. To retrieve the next set of models, use the token in the next request.

sortOrder :: Maybe OrderKey

The sort order for results. The default is Descending.

nameContains :: Maybe Text

A string in the model name. This filter returns only models whose name contains the specified string.

maxResults :: Maybe Natural

The maximum number of models to return in the response.

creationTimeBefore :: Maybe POSIX

A filter that returns only models created before the specified time (timestamp).

sortBy :: Maybe ModelSortKey

Sorts the list of results. The default is CreationTime.

creationTimeAfter :: Maybe POSIX

A filter that returns only models with a creation time greater than or equal to the specified time (timestamp).

newListModels :: ListModels

Create a value of ListModels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListModels', listModels_nextToken - If the response to a previous ListModels request was truncated, the response includes a NextToken. To retrieve the next set of models, use the token in the next request.

$sel:sortOrder:ListModels', listModels_sortOrder - The sort order for results. The default is Descending.

$sel:nameContains:ListModels', listModels_nameContains - A string in the model name. This filter returns only models whose name contains the specified string.

$sel:maxResults:ListModels', listModels_maxResults - The maximum number of models to return in the response.

$sel:creationTimeBefore:ListModels', listModels_creationTimeBefore - A filter that returns only models created before the specified time (timestamp).

$sel:sortBy:ListModels', listModels_sortBy - Sorts the list of results. The default is CreationTime.

$sel:creationTimeAfter:ListModels', listModels_creationTimeAfter - A filter that returns only models with a creation time greater than or equal to the specified time (timestamp).

Request Lenses

listModels_nextToken :: Lens' ListModels (Maybe Text)

If the response to a previous ListModels request was truncated, the response includes a NextToken. To retrieve the next set of models, use the token in the next request.

listModels_sortOrder :: Lens' ListModels (Maybe OrderKey)

The sort order for results. The default is Descending.

listModels_nameContains :: Lens' ListModels (Maybe Text)

A string in the model name. This filter returns only models whose name contains the specified string.

listModels_maxResults :: Lens' ListModels (Maybe Natural)

The maximum number of models to return in the response.

listModels_creationTimeBefore :: Lens' ListModels (Maybe UTCTime)

A filter that returns only models created before the specified time (timestamp).

listModels_sortBy :: Lens' ListModels (Maybe ModelSortKey)

Sorts the list of results. The default is CreationTime.

listModels_creationTimeAfter :: Lens' ListModels (Maybe UTCTime)

A filter that returns only models with a creation time greater than or equal to the specified time (timestamp).

Destructuring the Response

data ListModelsResponse

See: newListModelsResponse smart constructor.

Constructors

ListModelsResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of models, use it in the subsequent request.

httpStatus :: Int

The response's http status code.

models :: [ModelSummary]

An array of ModelSummary objects, each of which lists a model.

newListModelsResponse

Arguments

:: Int
$sel:httpStatus:ListModelsResponse'

-> ListModelsResponse

Create a value of ListModelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListModels', listModelsResponse_nextToken - If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of models, use it in the subsequent request.

$sel:httpStatus:ListModelsResponse', listModelsResponse_httpStatus - The response's http status code.

$sel:models:ListModelsResponse', listModelsResponse_models - An array of ModelSummary objects, each of which lists a model.

Response Lenses

listModelsResponse_nextToken :: Lens' ListModelsResponse (Maybe Text)

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of models, use it in the subsequent request.

listModelsResponse_httpStatus :: Lens' ListModelsResponse Int

The response's http status code.

listModelsResponse_models :: Lens' ListModelsResponse [ModelSummary]

An array of ModelSummary objects, each of which lists a model.