Network.AWS.SageMaker.ListMonitoringExecutions

Description

Returns list of all monitoring job executions.

This operation returns paginated results.

Creating a Request

data ListMonitoringExecutions

See: newListMonitoringExecutions smart constructor.

Constructors

ListMonitoringExecutions'

Fields

lastModifiedTimeBefore :: Maybe POSIX

A filter that returns only jobs modified after a specified time.

nextToken :: Maybe Text

The token returned if the response is truncated. To retrieve the next set of job executions, use it in the next request.

sortOrder :: Maybe SortOrder

Whether to sort the results in Ascending or Descending order. The default is Descending.

endpointName :: Maybe Text

Name of a specific endpoint to fetch jobs for.

monitoringJobDefinitionName :: Maybe Text

Gets a list of the monitoring job runs of the specified monitoring job definitions.

monitoringScheduleName :: Maybe Text

Name of a specific schedule to fetch jobs for.

maxResults :: Maybe Natural

The maximum number of jobs to return in the response. The default value is 10.

scheduledTimeAfter :: Maybe POSIX

Filter for jobs scheduled after a specified time.

creationTimeBefore :: Maybe POSIX

A filter that returns only jobs created before a specified time.

lastModifiedTimeAfter :: Maybe POSIX

A filter that returns only jobs modified before a specified time.

sortBy :: Maybe MonitoringExecutionSortKey

Whether to sort results by Status, CreationTime, ScheduledTime field. The default is CreationTime.

statusEquals :: Maybe ExecutionStatus

A filter that retrieves only jobs with a specific status.

monitoringTypeEquals :: Maybe MonitoringType

A filter that returns only the monitoring job runs of the specified monitoring type.

creationTimeAfter :: Maybe POSIX

A filter that returns only jobs created after a specified time.

scheduledTimeBefore :: Maybe POSIX

Filter for jobs scheduled before a specified time.

newListMonitoringExecutions :: ListMonitoringExecutions

Create a value of ListMonitoringExecutions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedTimeBefore:ListMonitoringExecutions', listMonitoringExecutions_lastModifiedTimeBefore - A filter that returns only jobs modified after a specified time.

$sel:nextToken:ListMonitoringExecutions', listMonitoringExecutions_nextToken - The token returned if the response is truncated. To retrieve the next set of job executions, use it in the next request.

$sel:sortOrder:ListMonitoringExecutions', listMonitoringExecutions_sortOrder - Whether to sort the results in Ascending or Descending order. The default is Descending.

$sel:endpointName:ListMonitoringExecutions', listMonitoringExecutions_endpointName - Name of a specific endpoint to fetch jobs for.

$sel:monitoringJobDefinitionName:ListMonitoringExecutions', listMonitoringExecutions_monitoringJobDefinitionName - Gets a list of the monitoring job runs of the specified monitoring job definitions.

$sel:monitoringScheduleName:ListMonitoringExecutions', listMonitoringExecutions_monitoringScheduleName - Name of a specific schedule to fetch jobs for.

$sel:maxResults:ListMonitoringExecutions', listMonitoringExecutions_maxResults - The maximum number of jobs to return in the response. The default value is 10.

$sel:scheduledTimeAfter:ListMonitoringExecutions', listMonitoringExecutions_scheduledTimeAfter - Filter for jobs scheduled after a specified time.

$sel:creationTimeBefore:ListMonitoringExecutions', listMonitoringExecutions_creationTimeBefore - A filter that returns only jobs created before a specified time.

$sel:lastModifiedTimeAfter:ListMonitoringExecutions', listMonitoringExecutions_lastModifiedTimeAfter - A filter that returns only jobs modified before a specified time.

$sel:sortBy:ListMonitoringExecutions', listMonitoringExecutions_sortBy - Whether to sort results by Status, CreationTime, ScheduledTime field. The default is CreationTime.

$sel:statusEquals:ListMonitoringExecutions', listMonitoringExecutions_statusEquals - A filter that retrieves only jobs with a specific status.

$sel:monitoringTypeEquals:ListMonitoringExecutions', listMonitoringExecutions_monitoringTypeEquals - A filter that returns only the monitoring job runs of the specified monitoring type.

$sel:creationTimeAfter:ListMonitoringExecutions', listMonitoringExecutions_creationTimeAfter - A filter that returns only jobs created after a specified time.

$sel:scheduledTimeBefore:ListMonitoringExecutions', listMonitoringExecutions_scheduledTimeBefore - Filter for jobs scheduled before a specified time.

Request Lenses

listMonitoringExecutions_lastModifiedTimeBefore :: Lens' ListMonitoringExecutions (Maybe UTCTime)

A filter that returns only jobs modified after a specified time.

listMonitoringExecutions_nextToken :: Lens' ListMonitoringExecutions (Maybe Text)

The token returned if the response is truncated. To retrieve the next set of job executions, use it in the next request.

listMonitoringExecutions_sortOrder :: Lens' ListMonitoringExecutions (Maybe SortOrder)

Whether to sort the results in Ascending or Descending order. The default is Descending.

listMonitoringExecutions_endpointName :: Lens' ListMonitoringExecutions (Maybe Text)

Name of a specific endpoint to fetch jobs for.

listMonitoringExecutions_monitoringJobDefinitionName :: Lens' ListMonitoringExecutions (Maybe Text)

Gets a list of the monitoring job runs of the specified monitoring job definitions.

listMonitoringExecutions_monitoringScheduleName :: Lens' ListMonitoringExecutions (Maybe Text)

Name of a specific schedule to fetch jobs for.

listMonitoringExecutions_maxResults :: Lens' ListMonitoringExecutions (Maybe Natural)

The maximum number of jobs to return in the response. The default value is 10.

listMonitoringExecutions_scheduledTimeAfter :: Lens' ListMonitoringExecutions (Maybe UTCTime)

Filter for jobs scheduled after a specified time.

listMonitoringExecutions_creationTimeBefore :: Lens' ListMonitoringExecutions (Maybe UTCTime)

A filter that returns only jobs created before a specified time.

listMonitoringExecutions_lastModifiedTimeAfter :: Lens' ListMonitoringExecutions (Maybe UTCTime)

A filter that returns only jobs modified before a specified time.

listMonitoringExecutions_sortBy :: Lens' ListMonitoringExecutions (Maybe MonitoringExecutionSortKey)

Whether to sort results by Status, CreationTime, ScheduledTime field. The default is CreationTime.

listMonitoringExecutions_statusEquals :: Lens' ListMonitoringExecutions (Maybe ExecutionStatus)

A filter that retrieves only jobs with a specific status.

listMonitoringExecutions_monitoringTypeEquals :: Lens' ListMonitoringExecutions (Maybe MonitoringType)

A filter that returns only the monitoring job runs of the specified monitoring type.

listMonitoringExecutions_creationTimeAfter :: Lens' ListMonitoringExecutions (Maybe UTCTime)

A filter that returns only jobs created after a specified time.

listMonitoringExecutions_scheduledTimeBefore :: Lens' ListMonitoringExecutions (Maybe UTCTime)

Filter for jobs scheduled before a specified time.

Destructuring the Response

data ListMonitoringExecutionsResponse

See: newListMonitoringExecutionsResponse smart constructor.

Constructors

ListMonitoringExecutionsResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of jobs, use it in the subsequent reques

httpStatus :: Int

The response's http status code.

monitoringExecutionSummaries :: [MonitoringExecutionSummary]

A JSON array in which each element is a summary for a monitoring execution.

newListMonitoringExecutionsResponse

Arguments

:: Int
$sel:httpStatus:ListMonitoringExecutionsResponse'

-> ListMonitoringExecutionsResponse

Create a value of ListMonitoringExecutionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMonitoringExecutions', listMonitoringExecutionsResponse_nextToken - If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of jobs, use it in the subsequent reques

$sel:httpStatus:ListMonitoringExecutionsResponse', listMonitoringExecutionsResponse_httpStatus - The response's http status code.

$sel:monitoringExecutionSummaries:ListMonitoringExecutionsResponse', listMonitoringExecutionsResponse_monitoringExecutionSummaries - A JSON array in which each element is a summary for a monitoring execution.

Response Lenses

listMonitoringExecutionsResponse_nextToken :: Lens' ListMonitoringExecutionsResponse (Maybe Text)

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of jobs, use it in the subsequent reques

listMonitoringExecutionsResponse_httpStatus :: Lens' ListMonitoringExecutionsResponse Int

The response's http status code.

listMonitoringExecutionsResponse_monitoringExecutionSummaries :: Lens' ListMonitoringExecutionsResponse [MonitoringExecutionSummary]

A JSON array in which each element is a summary for a monitoring execution.