Network.AWS.SageMaker.ListMonitoringSchedules

Description

Returns list of all monitoring schedules.

This operation returns paginated results.

Creating a Request

data ListMonitoringSchedules

See: newListMonitoringSchedules smart constructor.

Constructors

ListMonitoringSchedules'

Fields

nameContains :: Maybe Text

Filter for monitoring schedules whose name contains a specified string.

endpointName :: Maybe Text

Name of a specific endpoint to fetch schedules for.

lastModifiedTimeBefore :: Maybe POSIX

A filter that returns only monitoring schedules modified before a specified time.

creationTimeAfter :: Maybe POSIX

A filter that returns only monitoring schedules created after a specified time.

nextToken :: Maybe Text

The token returned if the response is truncated. To retrieve the next set of job executions, use it in the next request.

sortOrder :: Maybe SortOrder

Whether to sort the results in Ascending or Descending order. The default is Descending.

lastModifiedTimeAfter :: Maybe POSIX

A filter that returns only monitoring schedules modified after a specified time.

creationTimeBefore :: Maybe POSIX

A filter that returns only monitoring schedules created before a specified time.

statusEquals :: Maybe ScheduleStatus

A filter that returns only monitoring schedules modified before a specified time.

monitoringTypeEquals :: Maybe MonitoringType

A filter that returns only the monitoring schedules for the specified monitoring type.

maxResults :: Maybe Natural

The maximum number of jobs to return in the response. The default value is 10.

sortBy :: Maybe MonitoringScheduleSortKey

Whether to sort results by Status, CreationTime, ScheduledTime field. The default is CreationTime.

monitoringJobDefinitionName :: Maybe Text

Gets a list of the monitoring schedules for the specified monitoring job definition.

newListMonitoringSchedules :: ListMonitoringSchedules

Create a value of ListMonitoringSchedules with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nameContains:ListMonitoringSchedules', listMonitoringSchedules_nameContains - Filter for monitoring schedules whose name contains a specified string.

$sel:endpointName:ListMonitoringSchedules', listMonitoringSchedules_endpointName - Name of a specific endpoint to fetch schedules for.

$sel:lastModifiedTimeBefore:ListMonitoringSchedules', listMonitoringSchedules_lastModifiedTimeBefore - A filter that returns only monitoring schedules modified before a specified time.

$sel:creationTimeAfter:ListMonitoringSchedules', listMonitoringSchedules_creationTimeAfter - A filter that returns only monitoring schedules created after a specified time.

$sel:nextToken:ListMonitoringSchedules', listMonitoringSchedules_nextToken - The token returned if the response is truncated. To retrieve the next set of job executions, use it in the next request.

$sel:sortOrder:ListMonitoringSchedules', listMonitoringSchedules_sortOrder - Whether to sort the results in Ascending or Descending order. The default is Descending.

$sel:lastModifiedTimeAfter:ListMonitoringSchedules', listMonitoringSchedules_lastModifiedTimeAfter - A filter that returns only monitoring schedules modified after a specified time.

$sel:creationTimeBefore:ListMonitoringSchedules', listMonitoringSchedules_creationTimeBefore - A filter that returns only monitoring schedules created before a specified time.

$sel:statusEquals:ListMonitoringSchedules', listMonitoringSchedules_statusEquals - A filter that returns only monitoring schedules modified before a specified time.

$sel:monitoringTypeEquals:ListMonitoringSchedules', listMonitoringSchedules_monitoringTypeEquals - A filter that returns only the monitoring schedules for the specified monitoring type.

$sel:maxResults:ListMonitoringSchedules', listMonitoringSchedules_maxResults - The maximum number of jobs to return in the response. The default value is 10.

$sel:sortBy:ListMonitoringSchedules', listMonitoringSchedules_sortBy - Whether to sort results by Status, CreationTime, ScheduledTime field. The default is CreationTime.

$sel:monitoringJobDefinitionName:ListMonitoringSchedules', listMonitoringSchedules_monitoringJobDefinitionName - Gets a list of the monitoring schedules for the specified monitoring job definition.

Request Lenses

listMonitoringSchedules_nameContains :: Lens' ListMonitoringSchedules (Maybe Text)

Filter for monitoring schedules whose name contains a specified string.

listMonitoringSchedules_endpointName :: Lens' ListMonitoringSchedules (Maybe Text)

Name of a specific endpoint to fetch schedules for.

listMonitoringSchedules_lastModifiedTimeBefore :: Lens' ListMonitoringSchedules (Maybe UTCTime)

A filter that returns only monitoring schedules modified before a specified time.

listMonitoringSchedules_creationTimeAfter :: Lens' ListMonitoringSchedules (Maybe UTCTime)

A filter that returns only monitoring schedules created after a specified time.

listMonitoringSchedules_nextToken :: Lens' ListMonitoringSchedules (Maybe Text)

The token returned if the response is truncated. To retrieve the next set of job executions, use it in the next request.

listMonitoringSchedules_sortOrder :: Lens' ListMonitoringSchedules (Maybe SortOrder)

Whether to sort the results in Ascending or Descending order. The default is Descending.

listMonitoringSchedules_lastModifiedTimeAfter :: Lens' ListMonitoringSchedules (Maybe UTCTime)

A filter that returns only monitoring schedules modified after a specified time.

listMonitoringSchedules_creationTimeBefore :: Lens' ListMonitoringSchedules (Maybe UTCTime)

A filter that returns only monitoring schedules created before a specified time.

listMonitoringSchedules_statusEquals :: Lens' ListMonitoringSchedules (Maybe ScheduleStatus)

A filter that returns only monitoring schedules modified before a specified time.

listMonitoringSchedules_monitoringTypeEquals :: Lens' ListMonitoringSchedules (Maybe MonitoringType)

A filter that returns only the monitoring schedules for the specified monitoring type.

listMonitoringSchedules_maxResults :: Lens' ListMonitoringSchedules (Maybe Natural)

The maximum number of jobs to return in the response. The default value is 10.

listMonitoringSchedules_sortBy :: Lens' ListMonitoringSchedules (Maybe MonitoringScheduleSortKey)

Whether to sort results by Status, CreationTime, ScheduledTime field. The default is CreationTime.

listMonitoringSchedules_monitoringJobDefinitionName :: Lens' ListMonitoringSchedules (Maybe Text)

Gets a list of the monitoring schedules for the specified monitoring job definition.

Destructuring the Response

data ListMonitoringSchedulesResponse

See: newListMonitoringSchedulesResponse smart constructor.

Constructors

ListMonitoringSchedulesResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of jobs, use it in the subsequent request.

httpStatus :: Int

The response's http status code.

monitoringScheduleSummaries :: [MonitoringScheduleSummary]

A JSON array in which each element is a summary for a monitoring schedule.

newListMonitoringSchedulesResponse

Arguments

:: Int
$sel:httpStatus:ListMonitoringSchedulesResponse'

-> ListMonitoringSchedulesResponse

Create a value of ListMonitoringSchedulesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMonitoringSchedules', listMonitoringSchedulesResponse_nextToken - If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of jobs, use it in the subsequent request.

$sel:httpStatus:ListMonitoringSchedulesResponse', listMonitoringSchedulesResponse_httpStatus - The response's http status code.

$sel:monitoringScheduleSummaries:ListMonitoringSchedulesResponse', listMonitoringSchedulesResponse_monitoringScheduleSummaries - A JSON array in which each element is a summary for a monitoring schedule.

Response Lenses

listMonitoringSchedulesResponse_nextToken :: Lens' ListMonitoringSchedulesResponse (Maybe Text)

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of jobs, use it in the subsequent request.

listMonitoringSchedulesResponse_httpStatus :: Lens' ListMonitoringSchedulesResponse Int

The response's http status code.

listMonitoringSchedulesResponse_monitoringScheduleSummaries :: Lens' ListMonitoringSchedulesResponse [MonitoringScheduleSummary]

A JSON array in which each element is a summary for a monitoring schedule.