Network.AWS.SageMaker.ListNotebookInstanceLifecycleConfigs

Description

Lists notebook instance lifestyle configurations created with the CreateNotebookInstanceLifecycleConfig API.

This operation returns paginated results.

Creating a Request

data ListNotebookInstanceLifecycleConfigs

See: newListNotebookInstanceLifecycleConfigs smart constructor.

Constructors

ListNotebookInstanceLifecycleConfigs'

Fields

lastModifiedTimeBefore :: Maybe POSIX

A filter that returns only lifecycle configurations that were modified before the specified time (timestamp).

nextToken :: Maybe Text

If the result of a ListNotebookInstanceLifecycleConfigs request was truncated, the response includes a NextToken. To get the next set of lifecycle configurations, use the token in the next request.

sortOrder :: Maybe NotebookInstanceLifecycleConfigSortOrder

The sort order for results.

nameContains :: Maybe Text

A string in the lifecycle configuration name. This filter returns only lifecycle configurations whose name contains the specified string.

maxResults :: Maybe Natural

The maximum number of lifecycle configurations to return in the response.

creationTimeBefore :: Maybe POSIX

A filter that returns only lifecycle configurations that were created before the specified time (timestamp).

lastModifiedTimeAfter :: Maybe POSIX

A filter that returns only lifecycle configurations that were modified after the specified time (timestamp).

sortBy :: Maybe NotebookInstanceLifecycleConfigSortKey

Sorts the list of results. The default is CreationTime.

creationTimeAfter :: Maybe POSIX

A filter that returns only lifecycle configurations that were created after the specified time (timestamp).

newListNotebookInstanceLifecycleConfigs :: ListNotebookInstanceLifecycleConfigs

Create a value of ListNotebookInstanceLifecycleConfigs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedTimeBefore:ListNotebookInstanceLifecycleConfigs', listNotebookInstanceLifecycleConfigs_lastModifiedTimeBefore - A filter that returns only lifecycle configurations that were modified before the specified time (timestamp).

$sel:nextToken:ListNotebookInstanceLifecycleConfigs', listNotebookInstanceLifecycleConfigs_nextToken - If the result of a ListNotebookInstanceLifecycleConfigs request was truncated, the response includes a NextToken. To get the next set of lifecycle configurations, use the token in the next request.

$sel:sortOrder:ListNotebookInstanceLifecycleConfigs', listNotebookInstanceLifecycleConfigs_sortOrder - The sort order for results.

$sel:nameContains:ListNotebookInstanceLifecycleConfigs', listNotebookInstanceLifecycleConfigs_nameContains - A string in the lifecycle configuration name. This filter returns only lifecycle configurations whose name contains the specified string.

$sel:maxResults:ListNotebookInstanceLifecycleConfigs', listNotebookInstanceLifecycleConfigs_maxResults - The maximum number of lifecycle configurations to return in the response.

$sel:creationTimeBefore:ListNotebookInstanceLifecycleConfigs', listNotebookInstanceLifecycleConfigs_creationTimeBefore - A filter that returns only lifecycle configurations that were created before the specified time (timestamp).

$sel:lastModifiedTimeAfter:ListNotebookInstanceLifecycleConfigs', listNotebookInstanceLifecycleConfigs_lastModifiedTimeAfter - A filter that returns only lifecycle configurations that were modified after the specified time (timestamp).

$sel:sortBy:ListNotebookInstanceLifecycleConfigs', listNotebookInstanceLifecycleConfigs_sortBy - Sorts the list of results. The default is CreationTime.

$sel:creationTimeAfter:ListNotebookInstanceLifecycleConfigs', listNotebookInstanceLifecycleConfigs_creationTimeAfter - A filter that returns only lifecycle configurations that were created after the specified time (timestamp).

Request Lenses

listNotebookInstanceLifecycleConfigs_lastModifiedTimeBefore :: Lens' ListNotebookInstanceLifecycleConfigs (Maybe UTCTime)

A filter that returns only lifecycle configurations that were modified before the specified time (timestamp).

listNotebookInstanceLifecycleConfigs_nextToken :: Lens' ListNotebookInstanceLifecycleConfigs (Maybe Text)

If the result of a ListNotebookInstanceLifecycleConfigs request was truncated, the response includes a NextToken. To get the next set of lifecycle configurations, use the token in the next request.

listNotebookInstanceLifecycleConfigs_sortOrder :: Lens' ListNotebookInstanceLifecycleConfigs (Maybe NotebookInstanceLifecycleConfigSortOrder)

The sort order for results.

listNotebookInstanceLifecycleConfigs_nameContains :: Lens' ListNotebookInstanceLifecycleConfigs (Maybe Text)

A string in the lifecycle configuration name. This filter returns only lifecycle configurations whose name contains the specified string.

listNotebookInstanceLifecycleConfigs_maxResults :: Lens' ListNotebookInstanceLifecycleConfigs (Maybe Natural)

The maximum number of lifecycle configurations to return in the response.

listNotebookInstanceLifecycleConfigs_creationTimeBefore :: Lens' ListNotebookInstanceLifecycleConfigs (Maybe UTCTime)

A filter that returns only lifecycle configurations that were created before the specified time (timestamp).

listNotebookInstanceLifecycleConfigs_lastModifiedTimeAfter :: Lens' ListNotebookInstanceLifecycleConfigs (Maybe UTCTime)

A filter that returns only lifecycle configurations that were modified after the specified time (timestamp).

listNotebookInstanceLifecycleConfigs_sortBy :: Lens' ListNotebookInstanceLifecycleConfigs (Maybe NotebookInstanceLifecycleConfigSortKey)

Sorts the list of results. The default is CreationTime.

listNotebookInstanceLifecycleConfigs_creationTimeAfter :: Lens' ListNotebookInstanceLifecycleConfigs (Maybe UTCTime)

A filter that returns only lifecycle configurations that were created after the specified time (timestamp).

Destructuring the Response

data ListNotebookInstanceLifecycleConfigsResponse

See: newListNotebookInstanceLifecycleConfigsResponse smart constructor.

Constructors

ListNotebookInstanceLifecycleConfigsResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, Amazon SageMaker returns this token. To get the next set of lifecycle configurations, use it in the next request.

notebookInstanceLifecycleConfigs :: Maybe [NotebookInstanceLifecycleConfigSummary]

An array of NotebookInstanceLifecycleConfiguration objects, each listing a lifecycle configuration.

httpStatus :: Int

The response's http status code.

newListNotebookInstanceLifecycleConfigsResponse

Arguments

:: Int
$sel:httpStatus:ListNotebookInstanceLifecycleConfigsResponse'

-> ListNotebookInstanceLifecycleConfigsResponse

Create a value of ListNotebookInstanceLifecycleConfigsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListNotebookInstanceLifecycleConfigs', listNotebookInstanceLifecycleConfigsResponse_nextToken - If the response is truncated, Amazon SageMaker returns this token. To get the next set of lifecycle configurations, use it in the next request.

$sel:notebookInstanceLifecycleConfigs:ListNotebookInstanceLifecycleConfigsResponse', listNotebookInstanceLifecycleConfigsResponse_notebookInstanceLifecycleConfigs - An array of NotebookInstanceLifecycleConfiguration objects, each listing a lifecycle configuration.

$sel:httpStatus:ListNotebookInstanceLifecycleConfigsResponse', listNotebookInstanceLifecycleConfigsResponse_httpStatus - The response's http status code.

Response Lenses

listNotebookInstanceLifecycleConfigsResponse_nextToken :: Lens' ListNotebookInstanceLifecycleConfigsResponse (Maybe Text)

If the response is truncated, Amazon SageMaker returns this token. To get the next set of lifecycle configurations, use it in the next request.

listNotebookInstanceLifecycleConfigsResponse_notebookInstanceLifecycleConfigs :: Lens' ListNotebookInstanceLifecycleConfigsResponse (Maybe [NotebookInstanceLifecycleConfigSummary])

An array of NotebookInstanceLifecycleConfiguration objects, each listing a lifecycle configuration.

listNotebookInstanceLifecycleConfigsResponse_httpStatus :: Lens' ListNotebookInstanceLifecycleConfigsResponse Int

The response's http status code.