Network.AWS.SageMaker.ListNotebookInstances

Description

Returns a list of the Amazon SageMaker notebook instances in the requester's account in an Amazon Web Services Region.

This operation returns paginated results.

Creating a Request

data ListNotebookInstances

See: newListNotebookInstances smart constructor.

Constructors

ListNotebookInstances'

Fields

lastModifiedTimeBefore :: Maybe POSIX

A filter that returns only notebook instances that were modified before the specified time (timestamp).

nextToken :: Maybe Text

If the previous call to the ListNotebookInstances is truncated, the response includes a NextToken. You can use this token in your subsequent ListNotebookInstances request to fetch the next set of notebook instances.

You might specify a filter or a sort order in your request. When response is truncated, you must use the same values for the filer and sort order in the next request.

sortOrder :: Maybe NotebookInstanceSortOrder

The sort order for results.

nameContains :: Maybe Text

A string in the notebook instances' name. This filter returns only notebook instances whose name contains the specified string.

additionalCodeRepositoryEquals :: Maybe Text

A filter that returns only notebook instances with associated with the specified git repository.

maxResults :: Maybe Natural

The maximum number of notebook instances to return.

creationTimeBefore :: Maybe POSIX

A filter that returns only notebook instances that were created before the specified time (timestamp).

lastModifiedTimeAfter :: Maybe POSIX

A filter that returns only notebook instances that were modified after the specified time (timestamp).

defaultCodeRepositoryContains :: Maybe Text

A string in the name or URL of a Git repository associated with this notebook instance. This filter returns only notebook instances associated with a git repository with a name that contains the specified string.

sortBy :: Maybe NotebookInstanceSortKey

The field to sort results by. The default is Name.

statusEquals :: Maybe NotebookInstanceStatus

A filter that returns only notebook instances with the specified status.

notebookInstanceLifecycleConfigNameContains :: Maybe Text

A string in the name of a notebook instances lifecycle configuration associated with this notebook instance. This filter returns only notebook instances associated with a lifecycle configuration with a name that contains the specified string.

creationTimeAfter :: Maybe POSIX

A filter that returns only notebook instances that were created after the specified time (timestamp).

newListNotebookInstances :: ListNotebookInstances

Create a value of ListNotebookInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedTimeBefore:ListNotebookInstances', listNotebookInstances_lastModifiedTimeBefore - A filter that returns only notebook instances that were modified before the specified time (timestamp).

$sel:nextToken:ListNotebookInstances', listNotebookInstances_nextToken - If the previous call to the ListNotebookInstances is truncated, the response includes a NextToken. You can use this token in your subsequent ListNotebookInstances request to fetch the next set of notebook instances.

You might specify a filter or a sort order in your request. When response is truncated, you must use the same values for the filer and sort order in the next request.

$sel:sortOrder:ListNotebookInstances', listNotebookInstances_sortOrder - The sort order for results.

$sel:nameContains:ListNotebookInstances', listNotebookInstances_nameContains - A string in the notebook instances' name. This filter returns only notebook instances whose name contains the specified string.

$sel:additionalCodeRepositoryEquals:ListNotebookInstances', listNotebookInstances_additionalCodeRepositoryEquals - A filter that returns only notebook instances with associated with the specified git repository.

$sel:maxResults:ListNotebookInstances', listNotebookInstances_maxResults - The maximum number of notebook instances to return.

$sel:creationTimeBefore:ListNotebookInstances', listNotebookInstances_creationTimeBefore - A filter that returns only notebook instances that were created before the specified time (timestamp).

$sel:lastModifiedTimeAfter:ListNotebookInstances', listNotebookInstances_lastModifiedTimeAfter - A filter that returns only notebook instances that were modified after the specified time (timestamp).

$sel:defaultCodeRepositoryContains:ListNotebookInstances', listNotebookInstances_defaultCodeRepositoryContains - A string in the name or URL of a Git repository associated with this notebook instance. This filter returns only notebook instances associated with a git repository with a name that contains the specified string.

$sel:sortBy:ListNotebookInstances', listNotebookInstances_sortBy - The field to sort results by. The default is Name.

$sel:statusEquals:ListNotebookInstances', listNotebookInstances_statusEquals - A filter that returns only notebook instances with the specified status.

$sel:notebookInstanceLifecycleConfigNameContains:ListNotebookInstances', listNotebookInstances_notebookInstanceLifecycleConfigNameContains - A string in the name of a notebook instances lifecycle configuration associated with this notebook instance. This filter returns only notebook instances associated with a lifecycle configuration with a name that contains the specified string.

$sel:creationTimeAfter:ListNotebookInstances', listNotebookInstances_creationTimeAfter - A filter that returns only notebook instances that were created after the specified time (timestamp).

Request Lenses

listNotebookInstances_lastModifiedTimeBefore :: Lens' ListNotebookInstances (Maybe UTCTime)

A filter that returns only notebook instances that were modified before the specified time (timestamp).

listNotebookInstances_nextToken :: Lens' ListNotebookInstances (Maybe Text)

If the previous call to the ListNotebookInstances is truncated, the response includes a NextToken. You can use this token in your subsequent ListNotebookInstances request to fetch the next set of notebook instances.

You might specify a filter or a sort order in your request. When response is truncated, you must use the same values for the filer and sort order in the next request.

listNotebookInstances_sortOrder :: Lens' ListNotebookInstances (Maybe NotebookInstanceSortOrder)

The sort order for results.

listNotebookInstances_nameContains :: Lens' ListNotebookInstances (Maybe Text)

A string in the notebook instances' name. This filter returns only notebook instances whose name contains the specified string.

listNotebookInstances_additionalCodeRepositoryEquals :: Lens' ListNotebookInstances (Maybe Text)

A filter that returns only notebook instances with associated with the specified git repository.

listNotebookInstances_maxResults :: Lens' ListNotebookInstances (Maybe Natural)

The maximum number of notebook instances to return.

listNotebookInstances_creationTimeBefore :: Lens' ListNotebookInstances (Maybe UTCTime)

A filter that returns only notebook instances that were created before the specified time (timestamp).

listNotebookInstances_lastModifiedTimeAfter :: Lens' ListNotebookInstances (Maybe UTCTime)

A filter that returns only notebook instances that were modified after the specified time (timestamp).

listNotebookInstances_defaultCodeRepositoryContains :: Lens' ListNotebookInstances (Maybe Text)

A string in the name or URL of a Git repository associated with this notebook instance. This filter returns only notebook instances associated with a git repository with a name that contains the specified string.

listNotebookInstances_sortBy :: Lens' ListNotebookInstances (Maybe NotebookInstanceSortKey)

The field to sort results by. The default is Name.

listNotebookInstances_statusEquals :: Lens' ListNotebookInstances (Maybe NotebookInstanceStatus)

A filter that returns only notebook instances with the specified status.

listNotebookInstances_notebookInstanceLifecycleConfigNameContains :: Lens' ListNotebookInstances (Maybe Text)

A string in the name of a notebook instances lifecycle configuration associated with this notebook instance. This filter returns only notebook instances associated with a lifecycle configuration with a name that contains the specified string.

listNotebookInstances_creationTimeAfter :: Lens' ListNotebookInstances (Maybe UTCTime)

A filter that returns only notebook instances that were created after the specified time (timestamp).

Destructuring the Response

data ListNotebookInstancesResponse

See: newListNotebookInstancesResponse smart constructor.

Constructors

ListNotebookInstancesResponse'

Fields

nextToken :: Maybe Text

If the response to the previous ListNotebookInstances request was truncated, Amazon SageMaker returns this token. To retrieve the next set of notebook instances, use the token in the next request.

notebookInstances :: Maybe [NotebookInstanceSummary]

An array of NotebookInstanceSummary objects, one for each notebook instance.

httpStatus :: Int

The response's http status code.

newListNotebookInstancesResponse

Arguments

:: Int
$sel:httpStatus:ListNotebookInstancesResponse'

-> ListNotebookInstancesResponse

Create a value of ListNotebookInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListNotebookInstances', listNotebookInstancesResponse_nextToken - If the response to the previous ListNotebookInstances request was truncated, Amazon SageMaker returns this token. To retrieve the next set of notebook instances, use the token in the next request.

$sel:notebookInstances:ListNotebookInstancesResponse', listNotebookInstancesResponse_notebookInstances - An array of NotebookInstanceSummary objects, one for each notebook instance.

$sel:httpStatus:ListNotebookInstancesResponse', listNotebookInstancesResponse_httpStatus - The response's http status code.

Response Lenses

listNotebookInstancesResponse_nextToken :: Lens' ListNotebookInstancesResponse (Maybe Text)

If the response to the previous ListNotebookInstances request was truncated, Amazon SageMaker returns this token. To retrieve the next set of notebook instances, use the token in the next request.

listNotebookInstancesResponse_notebookInstances :: Lens' ListNotebookInstancesResponse (Maybe [NotebookInstanceSummary])

An array of NotebookInstanceSummary objects, one for each notebook instance.

listNotebookInstancesResponse_httpStatus :: Lens' ListNotebookInstancesResponse Int

The response's http status code.