Network.AWS.SageMaker.ListPipelineExecutionSteps

Description

Gets a list of PipeLineExecutionStep objects.

This operation returns paginated results.

Creating a Request

data ListPipelineExecutionSteps

See: newListPipelineExecutionSteps smart constructor.

Constructors

ListPipelineExecutionSteps'

Fields

nextToken :: Maybe Text

If the result of the previous ListPipelineExecutionSteps request was truncated, the response includes a NextToken. To retrieve the next set of pipeline execution steps, use the token in the next request.

sortOrder :: Maybe SortOrder

The field by which to sort results. The default is CreatedTime.

maxResults :: Maybe Natural

The maximum number of pipeline execution steps to return in the response.

pipelineExecutionArn :: Maybe Text

The Amazon Resource Name (ARN) of the pipeline execution.

newListPipelineExecutionSteps :: ListPipelineExecutionSteps

Create a value of ListPipelineExecutionSteps with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPipelineExecutionSteps', listPipelineExecutionSteps_nextToken - If the result of the previous ListPipelineExecutionSteps request was truncated, the response includes a NextToken. To retrieve the next set of pipeline execution steps, use the token in the next request.

$sel:sortOrder:ListPipelineExecutionSteps', listPipelineExecutionSteps_sortOrder - The field by which to sort results. The default is CreatedTime.

$sel:maxResults:ListPipelineExecutionSteps', listPipelineExecutionSteps_maxResults - The maximum number of pipeline execution steps to return in the response.

$sel:pipelineExecutionArn:ListPipelineExecutionSteps', listPipelineExecutionSteps_pipelineExecutionArn - The Amazon Resource Name (ARN) of the pipeline execution.

Request Lenses

listPipelineExecutionSteps_nextToken :: Lens' ListPipelineExecutionSteps (Maybe Text)

If the result of the previous ListPipelineExecutionSteps request was truncated, the response includes a NextToken. To retrieve the next set of pipeline execution steps, use the token in the next request.

listPipelineExecutionSteps_sortOrder :: Lens' ListPipelineExecutionSteps (Maybe SortOrder)

The field by which to sort results. The default is CreatedTime.

listPipelineExecutionSteps_maxResults :: Lens' ListPipelineExecutionSteps (Maybe Natural)

The maximum number of pipeline execution steps to return in the response.

listPipelineExecutionSteps_pipelineExecutionArn :: Lens' ListPipelineExecutionSteps (Maybe Text)

The Amazon Resource Name (ARN) of the pipeline execution.

Destructuring the Response

data ListPipelineExecutionStepsResponse

See: newListPipelineExecutionStepsResponse smart constructor.

Constructors

ListPipelineExecutionStepsResponse'

Fields

nextToken :: Maybe Text

If the result of the previous ListPipelineExecutionSteps request was truncated, the response includes a NextToken. To retrieve the next set of pipeline execution steps, use the token in the next request.

pipelineExecutionSteps :: Maybe [PipelineExecutionStep]

A list of PipeLineExecutionStep objects. Each PipeLineExecutionStep consists of StepName, StartTime, EndTime, StepStatus, and Metadata. Metadata is an object with properties for each job that contains relevant information about the job created by the step.

httpStatus :: Int

The response's http status code.

newListPipelineExecutionStepsResponse

Arguments

:: Int
$sel:httpStatus:ListPipelineExecutionStepsResponse'

-> ListPipelineExecutionStepsResponse

Create a value of ListPipelineExecutionStepsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPipelineExecutionSteps', listPipelineExecutionStepsResponse_nextToken - If the result of the previous ListPipelineExecutionSteps request was truncated, the response includes a NextToken. To retrieve the next set of pipeline execution steps, use the token in the next request.

$sel:pipelineExecutionSteps:ListPipelineExecutionStepsResponse', listPipelineExecutionStepsResponse_pipelineExecutionSteps - A list of PipeLineExecutionStep objects. Each PipeLineExecutionStep consists of StepName, StartTime, EndTime, StepStatus, and Metadata. Metadata is an object with properties for each job that contains relevant information about the job created by the step.

$sel:httpStatus:ListPipelineExecutionStepsResponse', listPipelineExecutionStepsResponse_httpStatus - The response's http status code.

Response Lenses

listPipelineExecutionStepsResponse_nextToken :: Lens' ListPipelineExecutionStepsResponse (Maybe Text)

If the result of the previous ListPipelineExecutionSteps request was truncated, the response includes a NextToken. To retrieve the next set of pipeline execution steps, use the token in the next request.

listPipelineExecutionStepsResponse_pipelineExecutionSteps :: Lens' ListPipelineExecutionStepsResponse (Maybe [PipelineExecutionStep])

A list of PipeLineExecutionStep objects. Each PipeLineExecutionStep consists of StepName, StartTime, EndTime, StepStatus, and Metadata. Metadata is an object with properties for each job that contains relevant information about the job created by the step.

listPipelineExecutionStepsResponse_httpStatus :: Lens' ListPipelineExecutionStepsResponse Int

The response's http status code.