Network.AWS.SageMaker.ListProjects

Description

Gets a list of the projects in an Amazon Web Services account.

Creating a Request

data ListProjects

See: newListProjects smart constructor.

Constructors

ListProjects'

Fields

nextToken :: Maybe Text

If the result of the previous ListProjects request was truncated, the response includes a NextToken. To retrieve the next set of projects, use the token in the next request.

sortOrder :: Maybe ProjectSortOrder

The sort order for results. The default is Ascending.

nameContains :: Maybe Text

A filter that returns the projects whose name contains a specified string.

maxResults :: Maybe Natural

The maximum number of projects to return in the response.

creationTimeBefore :: Maybe POSIX

A filter that returns the projects that were created before a specified time.

sortBy :: Maybe ProjectSortBy

The field by which to sort results. The default is CreationTime.

creationTimeAfter :: Maybe POSIX

A filter that returns the projects that were created after a specified time.

newListProjects :: ListProjects

Create a value of ListProjects with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListProjects', listProjects_nextToken - If the result of the previous ListProjects request was truncated, the response includes a NextToken. To retrieve the next set of projects, use the token in the next request.

$sel:sortOrder:ListProjects', listProjects_sortOrder - The sort order for results. The default is Ascending.

$sel:nameContains:ListProjects', listProjects_nameContains - A filter that returns the projects whose name contains a specified string.

$sel:maxResults:ListProjects', listProjects_maxResults - The maximum number of projects to return in the response.

$sel:creationTimeBefore:ListProjects', listProjects_creationTimeBefore - A filter that returns the projects that were created before a specified time.

$sel:sortBy:ListProjects', listProjects_sortBy - The field by which to sort results. The default is CreationTime.

$sel:creationTimeAfter:ListProjects', listProjects_creationTimeAfter - A filter that returns the projects that were created after a specified time.

Request Lenses

listProjects_nextToken :: Lens' ListProjects (Maybe Text)

If the result of the previous ListProjects request was truncated, the response includes a NextToken. To retrieve the next set of projects, use the token in the next request.

listProjects_sortOrder :: Lens' ListProjects (Maybe ProjectSortOrder)

The sort order for results. The default is Ascending.

listProjects_nameContains :: Lens' ListProjects (Maybe Text)

A filter that returns the projects whose name contains a specified string.

listProjects_maxResults :: Lens' ListProjects (Maybe Natural)

The maximum number of projects to return in the response.

listProjects_creationTimeBefore :: Lens' ListProjects (Maybe UTCTime)

A filter that returns the projects that were created before a specified time.

listProjects_sortBy :: Lens' ListProjects (Maybe ProjectSortBy)

The field by which to sort results. The default is CreationTime.

listProjects_creationTimeAfter :: Lens' ListProjects (Maybe UTCTime)

A filter that returns the projects that were created after a specified time.

Destructuring the Response

data ListProjectsResponse

See: newListProjectsResponse smart constructor.

Constructors

ListProjectsResponse'

Fields

nextToken :: Maybe Text

If the result of the previous ListCompilationJobs request was truncated, the response includes a NextToken. To retrieve the next set of model compilation jobs, use the token in the next request.

httpStatus :: Int

The response's http status code.

projectSummaryList :: [ProjectSummary]

A list of summaries of projects.

newListProjectsResponse

Arguments

:: Int
$sel:httpStatus:ListProjectsResponse'

-> ListProjectsResponse

Create a value of ListProjectsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListProjects', listProjectsResponse_nextToken - If the result of the previous ListCompilationJobs request was truncated, the response includes a NextToken. To retrieve the next set of model compilation jobs, use the token in the next request.

$sel:httpStatus:ListProjectsResponse', listProjectsResponse_httpStatus - The response's http status code.

$sel:projectSummaryList:ListProjectsResponse', listProjectsResponse_projectSummaryList - A list of summaries of projects.

Response Lenses

listProjectsResponse_nextToken :: Lens' ListProjectsResponse (Maybe Text)

If the result of the previous ListCompilationJobs request was truncated, the response includes a NextToken. To retrieve the next set of model compilation jobs, use the token in the next request.

listProjectsResponse_httpStatus :: Lens' ListProjectsResponse Int

The response's http status code.

listProjectsResponse_projectSummaryList :: Lens' ListProjectsResponse [ProjectSummary]

A list of summaries of projects.