Network.AWS.SageMaker.ListTrainingJobs

Description

Lists training jobs.

When StatusEquals and MaxResults are set at the same time, the MaxResults number of training jobs are first retrieved ignoring the StatusEquals parameter and then they are filtered by the StatusEquals parameter, which is returned as a response.

For example, if ListTrainingJobs is invoked with the following parameters:

{ ... MaxResults: 100, StatusEquals: InProgress ... }

First, 100 trainings jobs with any status, including those other than InProgress, are selected (sorted according to the creation time, from the most current to the oldest). Next, those with a status of InProgress are returned.

You can quickly test the API using the following Amazon Web Services CLI code.

aws sagemaker list-training-jobs --max-results 100 --status-equals InProgress

This operation returns paginated results.

Creating a Request

data ListTrainingJobs

See: newListTrainingJobs smart constructor.

Constructors

ListTrainingJobs'

Fields

lastModifiedTimeBefore :: Maybe POSIX

A filter that returns only training jobs modified before the specified time (timestamp).

nextToken :: Maybe Text

If the result of the previous ListTrainingJobs request was truncated, the response includes a NextToken. To retrieve the next set of training jobs, use the token in the next request.

sortOrder :: Maybe SortOrder

The sort order for results. The default is Ascending.

nameContains :: Maybe Text

A string in the training job name. This filter returns only training jobs whose name contains the specified string.

maxResults :: Maybe Natural

The maximum number of training jobs to return in the response.

creationTimeBefore :: Maybe POSIX

A filter that returns only training jobs created before the specified time (timestamp).

lastModifiedTimeAfter :: Maybe POSIX

A filter that returns only training jobs modified after the specified time (timestamp).

sortBy :: Maybe SortBy

The field to sort results by. The default is CreationTime.

statusEquals :: Maybe TrainingJobStatus

A filter that retrieves only training jobs with a specific status.

creationTimeAfter :: Maybe POSIX

A filter that returns only training jobs created after the specified time (timestamp).

newListTrainingJobs :: ListTrainingJobs

Create a value of ListTrainingJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedTimeBefore:ListTrainingJobs', listTrainingJobs_lastModifiedTimeBefore - A filter that returns only training jobs modified before the specified time (timestamp).

$sel:nextToken:ListTrainingJobs', listTrainingJobs_nextToken - If the result of the previous ListTrainingJobs request was truncated, the response includes a NextToken. To retrieve the next set of training jobs, use the token in the next request.

$sel:sortOrder:ListTrainingJobs', listTrainingJobs_sortOrder - The sort order for results. The default is Ascending.

$sel:nameContains:ListTrainingJobs', listTrainingJobs_nameContains - A string in the training job name. This filter returns only training jobs whose name contains the specified string.

$sel:maxResults:ListTrainingJobs', listTrainingJobs_maxResults - The maximum number of training jobs to return in the response.

$sel:creationTimeBefore:ListTrainingJobs', listTrainingJobs_creationTimeBefore - A filter that returns only training jobs created before the specified time (timestamp).

$sel:lastModifiedTimeAfter:ListTrainingJobs', listTrainingJobs_lastModifiedTimeAfter - A filter that returns only training jobs modified after the specified time (timestamp).

$sel:sortBy:ListTrainingJobs', listTrainingJobs_sortBy - The field to sort results by. The default is CreationTime.

$sel:statusEquals:ListTrainingJobs', listTrainingJobs_statusEquals - A filter that retrieves only training jobs with a specific status.

$sel:creationTimeAfter:ListTrainingJobs', listTrainingJobs_creationTimeAfter - A filter that returns only training jobs created after the specified time (timestamp).

Request Lenses

listTrainingJobs_lastModifiedTimeBefore :: Lens' ListTrainingJobs (Maybe UTCTime)

A filter that returns only training jobs modified before the specified time (timestamp).

listTrainingJobs_nextToken :: Lens' ListTrainingJobs (Maybe Text)

If the result of the previous ListTrainingJobs request was truncated, the response includes a NextToken. To retrieve the next set of training jobs, use the token in the next request.

listTrainingJobs_sortOrder :: Lens' ListTrainingJobs (Maybe SortOrder)

The sort order for results. The default is Ascending.

listTrainingJobs_nameContains :: Lens' ListTrainingJobs (Maybe Text)

A string in the training job name. This filter returns only training jobs whose name contains the specified string.

listTrainingJobs_maxResults :: Lens' ListTrainingJobs (Maybe Natural)

The maximum number of training jobs to return in the response.

listTrainingJobs_creationTimeBefore :: Lens' ListTrainingJobs (Maybe UTCTime)

A filter that returns only training jobs created before the specified time (timestamp).

listTrainingJobs_lastModifiedTimeAfter :: Lens' ListTrainingJobs (Maybe UTCTime)

A filter that returns only training jobs modified after the specified time (timestamp).

listTrainingJobs_sortBy :: Lens' ListTrainingJobs (Maybe SortBy)

The field to sort results by. The default is CreationTime.

listTrainingJobs_statusEquals :: Lens' ListTrainingJobs (Maybe TrainingJobStatus)

A filter that retrieves only training jobs with a specific status.

listTrainingJobs_creationTimeAfter :: Lens' ListTrainingJobs (Maybe UTCTime)

A filter that returns only training jobs created after the specified time (timestamp).

Destructuring the Response

data ListTrainingJobsResponse

See: newListTrainingJobsResponse smart constructor.

Constructors

ListTrainingJobsResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of training jobs, use it in the subsequent request.

httpStatus :: Int

The response's http status code.

trainingJobSummaries :: [TrainingJobSummary]

An array of TrainingJobSummary objects, each listing a training job.

newListTrainingJobsResponse

Arguments

:: Int
$sel:httpStatus:ListTrainingJobsResponse'

-> ListTrainingJobsResponse

Create a value of ListTrainingJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTrainingJobs', listTrainingJobsResponse_nextToken - If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of training jobs, use it in the subsequent request.

$sel:httpStatus:ListTrainingJobsResponse', listTrainingJobsResponse_httpStatus - The response's http status code.

$sel:trainingJobSummaries:ListTrainingJobsResponse', listTrainingJobsResponse_trainingJobSummaries - An array of TrainingJobSummary objects, each listing a training job.

Response Lenses

listTrainingJobsResponse_nextToken :: Lens' ListTrainingJobsResponse (Maybe Text)

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of training jobs, use it in the subsequent request.

listTrainingJobsResponse_httpStatus :: Lens' ListTrainingJobsResponse Int

The response's http status code.

listTrainingJobsResponse_trainingJobSummaries :: Lens' ListTrainingJobsResponse [TrainingJobSummary]

An array of TrainingJobSummary objects, each listing a training job.