Network.AWS.SageMaker.ListTrainingJobsForHyperParameterTuningJob

Description

Gets a list of TrainingJobSummary objects that describe the training jobs that a hyperparameter tuning job launched.

This operation returns paginated results.

Creating a Request

data ListTrainingJobsForHyperParameterTuningJob

See: newListTrainingJobsForHyperParameterTuningJob smart constructor.

Constructors

ListTrainingJobsForHyperParameterTuningJob'

Fields

nextToken :: Maybe Text

If the result of the previous ListTrainingJobsForHyperParameterTuningJob request was truncated, the response includes a NextToken. To retrieve the next set of training jobs, use the token in the next request.

sortOrder :: Maybe SortOrder

The sort order for results. The default is Ascending.

maxResults :: Maybe Natural

The maximum number of training jobs to return. The default value is 10.

sortBy :: Maybe TrainingJobSortByOptions

The field to sort results by. The default is Name.

If the value of this field is FinalObjectiveMetricValue, any training jobs that did not return an objective metric are not listed.

statusEquals :: Maybe TrainingJobStatus

A filter that returns only training jobs with the specified status.

hyperParameterTuningJobName :: Text

The name of the tuning job whose training jobs you want to list.

newListTrainingJobsForHyperParameterTuningJob

Arguments

:: Text
$sel:hyperParameterTuningJobName:ListTrainingJobsForHyperParameterTuningJob'

-> ListTrainingJobsForHyperParameterTuningJob

Create a value of ListTrainingJobsForHyperParameterTuningJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTrainingJobsForHyperParameterTuningJob', listTrainingJobsForHyperParameterTuningJob_nextToken - If the result of the previous ListTrainingJobsForHyperParameterTuningJob request was truncated, the response includes a NextToken. To retrieve the next set of training jobs, use the token in the next request.

$sel:sortOrder:ListTrainingJobsForHyperParameterTuningJob', listTrainingJobsForHyperParameterTuningJob_sortOrder - The sort order for results. The default is Ascending.

$sel:maxResults:ListTrainingJobsForHyperParameterTuningJob', listTrainingJobsForHyperParameterTuningJob_maxResults - The maximum number of training jobs to return. The default value is 10.

$sel:sortBy:ListTrainingJobsForHyperParameterTuningJob', listTrainingJobsForHyperParameterTuningJob_sortBy - The field to sort results by. The default is Name.

If the value of this field is FinalObjectiveMetricValue, any training jobs that did not return an objective metric are not listed.

$sel:statusEquals:ListTrainingJobsForHyperParameterTuningJob', listTrainingJobsForHyperParameterTuningJob_statusEquals - A filter that returns only training jobs with the specified status.

$sel:hyperParameterTuningJobName:ListTrainingJobsForHyperParameterTuningJob', listTrainingJobsForHyperParameterTuningJob_hyperParameterTuningJobName - The name of the tuning job whose training jobs you want to list.

Request Lenses

listTrainingJobsForHyperParameterTuningJob_nextToken :: Lens' ListTrainingJobsForHyperParameterTuningJob (Maybe Text)

If the result of the previous ListTrainingJobsForHyperParameterTuningJob request was truncated, the response includes a NextToken. To retrieve the next set of training jobs, use the token in the next request.

listTrainingJobsForHyperParameterTuningJob_sortOrder :: Lens' ListTrainingJobsForHyperParameterTuningJob (Maybe SortOrder)

The sort order for results. The default is Ascending.

listTrainingJobsForHyperParameterTuningJob_maxResults :: Lens' ListTrainingJobsForHyperParameterTuningJob (Maybe Natural)

The maximum number of training jobs to return. The default value is 10.

listTrainingJobsForHyperParameterTuningJob_sortBy :: Lens' ListTrainingJobsForHyperParameterTuningJob (Maybe TrainingJobSortByOptions)

The field to sort results by. The default is Name.

If the value of this field is FinalObjectiveMetricValue, any training jobs that did not return an objective metric are not listed.

listTrainingJobsForHyperParameterTuningJob_statusEquals :: Lens' ListTrainingJobsForHyperParameterTuningJob (Maybe TrainingJobStatus)

A filter that returns only training jobs with the specified status.

listTrainingJobsForHyperParameterTuningJob_hyperParameterTuningJobName :: Lens' ListTrainingJobsForHyperParameterTuningJob Text

The name of the tuning job whose training jobs you want to list.

Destructuring the Response

data ListTrainingJobsForHyperParameterTuningJobResponse

See: newListTrainingJobsForHyperParameterTuningJobResponse smart constructor.

Constructors

ListTrainingJobsForHyperParameterTuningJobResponse'

Fields

nextToken :: Maybe Text

If the result of this ListTrainingJobsForHyperParameterTuningJob request was truncated, the response includes a NextToken. To retrieve the next set of training jobs, use the token in the next request.

httpStatus :: Int

The response's http status code.

trainingJobSummaries :: [HyperParameterTrainingJobSummary]

A list of TrainingJobSummary objects that describe the training jobs that the ListTrainingJobsForHyperParameterTuningJob request returned.

newListTrainingJobsForHyperParameterTuningJobResponse

Arguments

:: Int
$sel:httpStatus:ListTrainingJobsForHyperParameterTuningJobResponse'

-> ListTrainingJobsForHyperParameterTuningJobResponse

Create a value of ListTrainingJobsForHyperParameterTuningJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTrainingJobsForHyperParameterTuningJob', listTrainingJobsForHyperParameterTuningJobResponse_nextToken - If the result of this ListTrainingJobsForHyperParameterTuningJob request was truncated, the response includes a NextToken. To retrieve the next set of training jobs, use the token in the next request.

$sel:httpStatus:ListTrainingJobsForHyperParameterTuningJobResponse', listTrainingJobsForHyperParameterTuningJobResponse_httpStatus - The response's http status code.

$sel:trainingJobSummaries:ListTrainingJobsForHyperParameterTuningJobResponse', listTrainingJobsForHyperParameterTuningJobResponse_trainingJobSummaries - A list of TrainingJobSummary objects that describe the training jobs that the ListTrainingJobsForHyperParameterTuningJob request returned.

Response Lenses

listTrainingJobsForHyperParameterTuningJobResponse_nextToken :: Lens' ListTrainingJobsForHyperParameterTuningJobResponse (Maybe Text)

If the result of this ListTrainingJobsForHyperParameterTuningJob request was truncated, the response includes a NextToken. To retrieve the next set of training jobs, use the token in the next request.

listTrainingJobsForHyperParameterTuningJobResponse_httpStatus :: Lens' ListTrainingJobsForHyperParameterTuningJobResponse Int

The response's http status code.

listTrainingJobsForHyperParameterTuningJobResponse_trainingJobSummaries :: Lens' ListTrainingJobsForHyperParameterTuningJobResponse [HyperParameterTrainingJobSummary]

A list of TrainingJobSummary objects that describe the training jobs that the ListTrainingJobsForHyperParameterTuningJob request returned.