Network.AWS.SageMaker.ListTransformJobs

Description

Lists transform jobs.

This operation returns paginated results.

Creating a Request

data ListTransformJobs

See: newListTransformJobs smart constructor.

Constructors

ListTransformJobs'

Fields

lastModifiedTimeBefore :: Maybe POSIX

A filter that returns only transform jobs modified before the specified time.

nextToken :: Maybe Text

If the result of the previous ListTransformJobs request was truncated, the response includes a NextToken. To retrieve the next set of transform jobs, use the token in the next request.

sortOrder :: Maybe SortOrder

The sort order for results. The default is Descending.

nameContains :: Maybe Text

A string in the transform job name. This filter returns only transform jobs whose name contains the specified string.

maxResults :: Maybe Natural

The maximum number of transform jobs to return in the response. The default value is 10.

creationTimeBefore :: Maybe POSIX

A filter that returns only transform jobs created before the specified time.

lastModifiedTimeAfter :: Maybe POSIX

A filter that returns only transform jobs modified after the specified time.

sortBy :: Maybe SortBy

The field to sort results by. The default is CreationTime.

statusEquals :: Maybe TransformJobStatus

A filter that retrieves only transform jobs with a specific status.

creationTimeAfter :: Maybe POSIX

A filter that returns only transform jobs created after the specified time.

newListTransformJobs :: ListTransformJobs

Create a value of ListTransformJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedTimeBefore:ListTransformJobs', listTransformJobs_lastModifiedTimeBefore - A filter that returns only transform jobs modified before the specified time.

$sel:nextToken:ListTransformJobs', listTransformJobs_nextToken - If the result of the previous ListTransformJobs request was truncated, the response includes a NextToken. To retrieve the next set of transform jobs, use the token in the next request.

$sel:sortOrder:ListTransformJobs', listTransformJobs_sortOrder - The sort order for results. The default is Descending.

$sel:nameContains:ListTransformJobs', listTransformJobs_nameContains - A string in the transform job name. This filter returns only transform jobs whose name contains the specified string.

$sel:maxResults:ListTransformJobs', listTransformJobs_maxResults - The maximum number of transform jobs to return in the response. The default value is 10.

$sel:creationTimeBefore:ListTransformJobs', listTransformJobs_creationTimeBefore - A filter that returns only transform jobs created before the specified time.

$sel:lastModifiedTimeAfter:ListTransformJobs', listTransformJobs_lastModifiedTimeAfter - A filter that returns only transform jobs modified after the specified time.

$sel:sortBy:ListTransformJobs', listTransformJobs_sortBy - The field to sort results by. The default is CreationTime.

$sel:statusEquals:ListTransformJobs', listTransformJobs_statusEquals - A filter that retrieves only transform jobs with a specific status.

$sel:creationTimeAfter:ListTransformJobs', listTransformJobs_creationTimeAfter - A filter that returns only transform jobs created after the specified time.

Request Lenses

listTransformJobs_lastModifiedTimeBefore :: Lens' ListTransformJobs (Maybe UTCTime)

A filter that returns only transform jobs modified before the specified time.

listTransformJobs_nextToken :: Lens' ListTransformJobs (Maybe Text)

If the result of the previous ListTransformJobs request was truncated, the response includes a NextToken. To retrieve the next set of transform jobs, use the token in the next request.

listTransformJobs_sortOrder :: Lens' ListTransformJobs (Maybe SortOrder)

The sort order for results. The default is Descending.

listTransformJobs_nameContains :: Lens' ListTransformJobs (Maybe Text)

A string in the transform job name. This filter returns only transform jobs whose name contains the specified string.

listTransformJobs_maxResults :: Lens' ListTransformJobs (Maybe Natural)

The maximum number of transform jobs to return in the response. The default value is 10.

listTransformJobs_creationTimeBefore :: Lens' ListTransformJobs (Maybe UTCTime)

A filter that returns only transform jobs created before the specified time.

listTransformJobs_lastModifiedTimeAfter :: Lens' ListTransformJobs (Maybe UTCTime)

A filter that returns only transform jobs modified after the specified time.

listTransformJobs_sortBy :: Lens' ListTransformJobs (Maybe SortBy)

The field to sort results by. The default is CreationTime.

listTransformJobs_statusEquals :: Lens' ListTransformJobs (Maybe TransformJobStatus)

A filter that retrieves only transform jobs with a specific status.

listTransformJobs_creationTimeAfter :: Lens' ListTransformJobs (Maybe UTCTime)

A filter that returns only transform jobs created after the specified time.

Destructuring the Response

data ListTransformJobsResponse

See: newListTransformJobsResponse smart constructor.

Constructors

ListTransformJobsResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of transform jobs, use it in the next request.

httpStatus :: Int

The response's http status code.

transformJobSummaries :: [TransformJobSummary]

An array of TransformJobSummary objects.

newListTransformJobsResponse

Arguments

:: Int
$sel:httpStatus:ListTransformJobsResponse'

-> ListTransformJobsResponse

Create a value of ListTransformJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTransformJobs', listTransformJobsResponse_nextToken - If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of transform jobs, use it in the next request.

$sel:httpStatus:ListTransformJobsResponse', listTransformJobsResponse_httpStatus - The response's http status code.

$sel:transformJobSummaries:ListTransformJobsResponse', listTransformJobsResponse_transformJobSummaries - An array of TransformJobSummary objects.

Response Lenses

listTransformJobsResponse_nextToken :: Lens' ListTransformJobsResponse (Maybe Text)

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of transform jobs, use it in the next request.

listTransformJobsResponse_httpStatus :: Lens' ListTransformJobsResponse Int

The response's http status code.

listTransformJobsResponse_transformJobSummaries :: Lens' ListTransformJobsResponse [TransformJobSummary]

An array of TransformJobSummary objects.