Network.AWS.SageMaker.ListTrialComponents

Description

Lists the trial components in your account. You can sort the list by trial component name or creation time. You can filter the list to show only components that were created in a specific time range. You can also filter on one of the following:

ExperimentName

SourceArn

TrialName

This operation returns paginated results.

Creating a Request

data ListTrialComponents

See: newListTrialComponents smart constructor.

Constructors

ListTrialComponents'

Fields

nextToken :: Maybe Text

If the previous call to ListTrialComponents didn't return the full set of components, the call returns a token for getting the next set of components.

sortOrder :: Maybe SortOrder

The sort order. The default value is Descending.

createdAfter :: Maybe POSIX

A filter that returns only components created after the specified time.

createdBefore :: Maybe POSIX

A filter that returns only components created before the specified time.

maxResults :: Maybe Natural

The maximum number of components to return in the response. The default value is 10.

sortBy :: Maybe SortTrialComponentsBy

The property used to sort results. The default value is CreationTime.

experimentName :: Maybe Text

A filter that returns only components that are part of the specified experiment. If you specify ExperimentName, you can't filter by SourceArn or TrialName.

sourceArn :: Maybe Text

A filter that returns only components that have the specified source Amazon Resource Name (ARN). If you specify SourceArn, you can't filter by ExperimentName or TrialName.

trialName :: Maybe Text

A filter that returns only components that are part of the specified trial. If you specify TrialName, you can't filter by ExperimentName or SourceArn.

newListTrialComponents :: ListTrialComponents

Create a value of ListTrialComponents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTrialComponents', listTrialComponents_nextToken - If the previous call to ListTrialComponents didn't return the full set of components, the call returns a token for getting the next set of components.

$sel:sortOrder:ListTrialComponents', listTrialComponents_sortOrder - The sort order. The default value is Descending.

$sel:createdAfter:ListTrialComponents', listTrialComponents_createdAfter - A filter that returns only components created after the specified time.

$sel:createdBefore:ListTrialComponents', listTrialComponents_createdBefore - A filter that returns only components created before the specified time.

$sel:maxResults:ListTrialComponents', listTrialComponents_maxResults - The maximum number of components to return in the response. The default value is 10.

$sel:sortBy:ListTrialComponents', listTrialComponents_sortBy - The property used to sort results. The default value is CreationTime.

$sel:experimentName:ListTrialComponents', listTrialComponents_experimentName - A filter that returns only components that are part of the specified experiment. If you specify ExperimentName, you can't filter by SourceArn or TrialName.

$sel:sourceArn:ListTrialComponents', listTrialComponents_sourceArn - A filter that returns only components that have the specified source Amazon Resource Name (ARN). If you specify SourceArn, you can't filter by ExperimentName or TrialName.

$sel:trialName:ListTrialComponents', listTrialComponents_trialName - A filter that returns only components that are part of the specified trial. If you specify TrialName, you can't filter by ExperimentName or SourceArn.

Request Lenses

listTrialComponents_nextToken :: Lens' ListTrialComponents (Maybe Text)

If the previous call to ListTrialComponents didn't return the full set of components, the call returns a token for getting the next set of components.

listTrialComponents_sortOrder :: Lens' ListTrialComponents (Maybe SortOrder)

The sort order. The default value is Descending.

listTrialComponents_createdAfter :: Lens' ListTrialComponents (Maybe UTCTime)

A filter that returns only components created after the specified time.

listTrialComponents_createdBefore :: Lens' ListTrialComponents (Maybe UTCTime)

A filter that returns only components created before the specified time.

listTrialComponents_maxResults :: Lens' ListTrialComponents (Maybe Natural)

The maximum number of components to return in the response. The default value is 10.

listTrialComponents_sortBy :: Lens' ListTrialComponents (Maybe SortTrialComponentsBy)

The property used to sort results. The default value is CreationTime.

listTrialComponents_experimentName :: Lens' ListTrialComponents (Maybe Text)

A filter that returns only components that are part of the specified experiment. If you specify ExperimentName, you can't filter by SourceArn or TrialName.

listTrialComponents_sourceArn :: Lens' ListTrialComponents (Maybe Text)

A filter that returns only components that have the specified source Amazon Resource Name (ARN). If you specify SourceArn, you can't filter by ExperimentName or TrialName.

listTrialComponents_trialName :: Lens' ListTrialComponents (Maybe Text)

A filter that returns only components that are part of the specified trial. If you specify TrialName, you can't filter by ExperimentName or SourceArn.

Destructuring the Response

data ListTrialComponentsResponse

See: newListTrialComponentsResponse smart constructor.

Constructors

ListTrialComponentsResponse'

Fields

nextToken :: Maybe Text

A token for getting the next set of components, if there are any.

trialComponentSummaries :: Maybe [TrialComponentSummary]

A list of the summaries of your trial components.

httpStatus :: Int

The response's http status code.

newListTrialComponentsResponse

Arguments

:: Int
$sel:httpStatus:ListTrialComponentsResponse'

-> ListTrialComponentsResponse

Create a value of ListTrialComponentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTrialComponents', listTrialComponentsResponse_nextToken - A token for getting the next set of components, if there are any.

$sel:trialComponentSummaries:ListTrialComponentsResponse', listTrialComponentsResponse_trialComponentSummaries - A list of the summaries of your trial components.

$sel:httpStatus:ListTrialComponentsResponse', listTrialComponentsResponse_httpStatus - The response's http status code.

Response Lenses

listTrialComponentsResponse_nextToken :: Lens' ListTrialComponentsResponse (Maybe Text)

A token for getting the next set of components, if there are any.

listTrialComponentsResponse_trialComponentSummaries :: Lens' ListTrialComponentsResponse (Maybe [TrialComponentSummary])

A list of the summaries of your trial components.

listTrialComponentsResponse_httpStatus :: Lens' ListTrialComponentsResponse Int

The response's http status code.