Network.AWS.SageMaker.ListUserProfiles

Description

Lists user profiles.

This operation returns paginated results.

Creating a Request

data ListUserProfiles

See: newListUserProfiles smart constructor.

Constructors

ListUserProfiles'

Fields

nextToken :: Maybe Text

If the previous response was truncated, you will receive this token. Use it in your next request to receive the next set of results.

sortOrder :: Maybe SortOrder

The sort order for the results. The default is Ascending.

userProfileNameContains :: Maybe Text

A parameter by which to filter the results.

maxResults :: Maybe Natural

Returns a list up to a specified limit.

domainIdEquals :: Maybe Text

A parameter by which to filter the results.

sortBy :: Maybe UserProfileSortKey

The parameter by which to sort the results. The default is CreationTime.

newListUserProfiles :: ListUserProfiles

Create a value of ListUserProfiles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListUserProfiles', listUserProfiles_nextToken - If the previous response was truncated, you will receive this token. Use it in your next request to receive the next set of results.

$sel:sortOrder:ListUserProfiles', listUserProfiles_sortOrder - The sort order for the results. The default is Ascending.

$sel:userProfileNameContains:ListUserProfiles', listUserProfiles_userProfileNameContains - A parameter by which to filter the results.

$sel:maxResults:ListUserProfiles', listUserProfiles_maxResults - Returns a list up to a specified limit.

$sel:domainIdEquals:ListUserProfiles', listUserProfiles_domainIdEquals - A parameter by which to filter the results.

$sel:sortBy:ListUserProfiles', listUserProfiles_sortBy - The parameter by which to sort the results. The default is CreationTime.

Request Lenses

listUserProfiles_nextToken :: Lens' ListUserProfiles (Maybe Text)

If the previous response was truncated, you will receive this token. Use it in your next request to receive the next set of results.

listUserProfiles_sortOrder :: Lens' ListUserProfiles (Maybe SortOrder)

The sort order for the results. The default is Ascending.

listUserProfiles_userProfileNameContains :: Lens' ListUserProfiles (Maybe Text)

A parameter by which to filter the results.

listUserProfiles_maxResults :: Lens' ListUserProfiles (Maybe Natural)

Returns a list up to a specified limit.

listUserProfiles_domainIdEquals :: Lens' ListUserProfiles (Maybe Text)

A parameter by which to filter the results.

listUserProfiles_sortBy :: Lens' ListUserProfiles (Maybe UserProfileSortKey)

The parameter by which to sort the results. The default is CreationTime.

Destructuring the Response

data ListUserProfilesResponse

See: newListUserProfilesResponse smart constructor.

Constructors

ListUserProfilesResponse'

Fields

nextToken :: Maybe Text

If the previous response was truncated, you will receive this token. Use it in your next request to receive the next set of results.

userProfiles :: Maybe [UserProfileDetails]

The list of user profiles.

httpStatus :: Int

The response's http status code.

newListUserProfilesResponse

Arguments

:: Int
$sel:httpStatus:ListUserProfilesResponse'

-> ListUserProfilesResponse

Create a value of ListUserProfilesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListUserProfiles', listUserProfilesResponse_nextToken - If the previous response was truncated, you will receive this token. Use it in your next request to receive the next set of results.

$sel:userProfiles:ListUserProfilesResponse', listUserProfilesResponse_userProfiles - The list of user profiles.

$sel:httpStatus:ListUserProfilesResponse', listUserProfilesResponse_httpStatus - The response's http status code.

Response Lenses

listUserProfilesResponse_nextToken :: Lens' ListUserProfilesResponse (Maybe Text)

If the previous response was truncated, you will receive this token. Use it in your next request to receive the next set of results.

listUserProfilesResponse_userProfiles :: Lens' ListUserProfilesResponse (Maybe [UserProfileDetails])

The list of user profiles.

listUserProfilesResponse_httpStatus :: Lens' ListUserProfilesResponse Int

The response's http status code.