Network.AWS.SageMaker.ListWorkteams

Description

Gets a list of private work teams that you have defined in a region. The list may be empty if no work team satisfies the filter specified in the NameContains parameter.

This operation returns paginated results.

Creating a Request

data ListWorkteams

See: newListWorkteams smart constructor.

Constructors

ListWorkteams'

Fields

nextToken :: Maybe Text

If the result of the previous ListWorkteams request was truncated, the response includes a NextToken. To retrieve the next set of labeling jobs, use the token in the next request.

sortOrder :: Maybe SortOrder

The sort order for results. The default is Ascending.

nameContains :: Maybe Text

A string in the work team's name. This filter returns only work teams whose name contains the specified string.

maxResults :: Maybe Natural

The maximum number of work teams to return in each page of the response.

sortBy :: Maybe ListWorkteamsSortByOptions

The field to sort results by. The default is CreationTime.

newListWorkteams :: ListWorkteams

Create a value of ListWorkteams with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListWorkteams', listWorkteams_nextToken - If the result of the previous ListWorkteams request was truncated, the response includes a NextToken. To retrieve the next set of labeling jobs, use the token in the next request.

$sel:sortOrder:ListWorkteams', listWorkteams_sortOrder - The sort order for results. The default is Ascending.

$sel:nameContains:ListWorkteams', listWorkteams_nameContains - A string in the work team's name. This filter returns only work teams whose name contains the specified string.

$sel:maxResults:ListWorkteams', listWorkteams_maxResults - The maximum number of work teams to return in each page of the response.

$sel:sortBy:ListWorkteams', listWorkteams_sortBy - The field to sort results by. The default is CreationTime.

Request Lenses

listWorkteams_nextToken :: Lens' ListWorkteams (Maybe Text)

If the result of the previous ListWorkteams request was truncated, the response includes a NextToken. To retrieve the next set of labeling jobs, use the token in the next request.

listWorkteams_sortOrder :: Lens' ListWorkteams (Maybe SortOrder)

The sort order for results. The default is Ascending.

listWorkteams_nameContains :: Lens' ListWorkteams (Maybe Text)

A string in the work team's name. This filter returns only work teams whose name contains the specified string.

listWorkteams_maxResults :: Lens' ListWorkteams (Maybe Natural)

The maximum number of work teams to return in each page of the response.

listWorkteams_sortBy :: Lens' ListWorkteams (Maybe ListWorkteamsSortByOptions)

The field to sort results by. The default is CreationTime.

Destructuring the Response

data ListWorkteamsResponse

See: newListWorkteamsResponse smart constructor.

Constructors

ListWorkteamsResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of work teams, use it in the subsequent request.

httpStatus :: Int

The response's http status code.

workteams :: [Workteam]

An array of Workteam objects, each describing a work team.

newListWorkteamsResponse

Arguments

:: Int
$sel:httpStatus:ListWorkteamsResponse'

-> ListWorkteamsResponse

Create a value of ListWorkteamsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListWorkteams', listWorkteamsResponse_nextToken - If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of work teams, use it in the subsequent request.

$sel:httpStatus:ListWorkteamsResponse', listWorkteamsResponse_httpStatus - The response's http status code.

$sel:workteams:ListWorkteamsResponse', listWorkteamsResponse_workteams - An array of Workteam objects, each describing a work team.

Response Lenses

listWorkteamsResponse_nextToken :: Lens' ListWorkteamsResponse (Maybe Text)

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of work teams, use it in the subsequent request.

listWorkteamsResponse_httpStatus :: Lens' ListWorkteamsResponse Int

The response's http status code.

listWorkteamsResponse_workteams :: Lens' ListWorkteamsResponse [Workteam]

An array of Workteam objects, each describing a work team.