Network.AWS.SageMaker.RetryPipelineExecution

Description

Retry the execution of the pipeline.

Creating a Request

data RetryPipelineExecution

See: newRetryPipelineExecution smart constructor.

Constructors

RetryPipelineExecution'

Fields

pipelineExecutionArn :: Text

The Amazon Resource Name (ARN) of the pipeline execution.

clientRequestToken :: Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the operation. An idempotent operation completes no more than once.

newRetryPipelineExecution

Arguments

:: Text
$sel:pipelineExecutionArn:RetryPipelineExecution'

-> Text
$sel:clientRequestToken:RetryPipelineExecution'

-> RetryPipelineExecution

Create a value of RetryPipelineExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineExecutionArn:RetryPipelineExecution', retryPipelineExecution_pipelineExecutionArn - The Amazon Resource Name (ARN) of the pipeline execution.

$sel:clientRequestToken:RetryPipelineExecution', retryPipelineExecution_clientRequestToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the operation. An idempotent operation completes no more than once.

Request Lenses

retryPipelineExecution_pipelineExecutionArn :: Lens' RetryPipelineExecution Text

The Amazon Resource Name (ARN) of the pipeline execution.

retryPipelineExecution_clientRequestToken :: Lens' RetryPipelineExecution Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the operation. An idempotent operation completes no more than once.

Destructuring the Response

data RetryPipelineExecutionResponse

See: newRetryPipelineExecutionResponse smart constructor.

Constructors

RetryPipelineExecutionResponse'

Fields

pipelineExecutionArn :: Maybe Text

The Amazon Resource Name (ARN) of the pipeline execution.

httpStatus :: Int

The response's http status code.

newRetryPipelineExecutionResponse

Arguments

:: Int
$sel:httpStatus:RetryPipelineExecutionResponse'

-> RetryPipelineExecutionResponse

Create a value of RetryPipelineExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineExecutionArn:RetryPipelineExecution', retryPipelineExecutionResponse_pipelineExecutionArn - The Amazon Resource Name (ARN) of the pipeline execution.

$sel:httpStatus:RetryPipelineExecutionResponse', retryPipelineExecutionResponse_httpStatus - The response's http status code.

Response Lenses

retryPipelineExecutionResponse_pipelineExecutionArn :: Lens' RetryPipelineExecutionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the pipeline execution.

retryPipelineExecutionResponse_httpStatus :: Lens' RetryPipelineExecutionResponse Int

The response's http status code.