Network.AWS.SageMaker.Types.AlgorithmSummary

Description

Documentation

data AlgorithmSummary

Provides summary information about an algorithm.

See: newAlgorithmSummary smart constructor.

Constructors

AlgorithmSummary'

Fields

algorithmDescription :: Maybe Text

A brief description of the algorithm.

algorithmName :: Text

The name of the algorithm that is described by the summary.

algorithmArn :: Text

The Amazon Resource Name (ARN) of the algorithm.

creationTime :: POSIX

A timestamp that shows when the algorithm was created.

algorithmStatus :: AlgorithmStatus

The overall status of the algorithm.

newAlgorithmSummary

Arguments

:: Text
$sel:algorithmName:AlgorithmSummary'

-> Text
$sel:algorithmArn:AlgorithmSummary'

-> UTCTime
$sel:creationTime:AlgorithmSummary'

-> AlgorithmStatus
$sel:algorithmStatus:AlgorithmSummary'

-> AlgorithmSummary

Create a value of AlgorithmSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:algorithmDescription:AlgorithmSummary', algorithmSummary_algorithmDescription - A brief description of the algorithm.

$sel:algorithmName:AlgorithmSummary', algorithmSummary_algorithmName - The name of the algorithm that is described by the summary.

$sel:algorithmArn:AlgorithmSummary', algorithmSummary_algorithmArn - The Amazon Resource Name (ARN) of the algorithm.

$sel:creationTime:AlgorithmSummary', algorithmSummary_creationTime - A timestamp that shows when the algorithm was created.

$sel:algorithmStatus:AlgorithmSummary', algorithmSummary_algorithmStatus - The overall status of the algorithm.

algorithmSummary_algorithmDescription :: Lens' AlgorithmSummary (Maybe Text)

A brief description of the algorithm.

algorithmSummary_algorithmName :: Lens' AlgorithmSummary Text

The name of the algorithm that is described by the summary.

algorithmSummary_algorithmArn :: Lens' AlgorithmSummary Text

The Amazon Resource Name (ARN) of the algorithm.

algorithmSummary_creationTime :: Lens' AlgorithmSummary UTCTime

A timestamp that shows when the algorithm was created.

algorithmSummary_algorithmStatus :: Lens' AlgorithmSummary AlgorithmStatus

The overall status of the algorithm.