Network.AWS.SageMaker.Types.AppImageConfigDetails

Description

Documentation

data AppImageConfigDetails

The configuration for running a SageMaker image as a KernelGateway app.

See: newAppImageConfigDetails smart constructor.

Constructors

AppImageConfigDetails'

Fields

creationTime :: Maybe POSIX

When the AppImageConfig was created.

appImageConfigArn :: Maybe Text

The Amazon Resource Name (ARN) of the AppImageConfig.

kernelGatewayImageConfig :: Maybe KernelGatewayImageConfig

The configuration for the file system and kernels in the SageMaker image.

appImageConfigName :: Maybe Text

The name of the AppImageConfig. Must be unique to your account.

lastModifiedTime :: Maybe POSIX

When the AppImageConfig was last modified.

newAppImageConfigDetails :: AppImageConfigDetails

Create a value of AppImageConfigDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:AppImageConfigDetails', appImageConfigDetails_creationTime - When the AppImageConfig was created.

$sel:appImageConfigArn:AppImageConfigDetails', appImageConfigDetails_appImageConfigArn - The Amazon Resource Name (ARN) of the AppImageConfig.

$sel:kernelGatewayImageConfig:AppImageConfigDetails', appImageConfigDetails_kernelGatewayImageConfig - The configuration for the file system and kernels in the SageMaker image.

$sel:appImageConfigName:AppImageConfigDetails', appImageConfigDetails_appImageConfigName - The name of the AppImageConfig. Must be unique to your account.

$sel:lastModifiedTime:AppImageConfigDetails', appImageConfigDetails_lastModifiedTime - When the AppImageConfig was last modified.

appImageConfigDetails_creationTime :: Lens' AppImageConfigDetails (Maybe UTCTime)

When the AppImageConfig was created.

appImageConfigDetails_appImageConfigArn :: Lens' AppImageConfigDetails (Maybe Text)

The Amazon Resource Name (ARN) of the AppImageConfig.

appImageConfigDetails_kernelGatewayImageConfig :: Lens' AppImageConfigDetails (Maybe KernelGatewayImageConfig)

The configuration for the file system and kernels in the SageMaker image.

appImageConfigDetails_appImageConfigName :: Lens' AppImageConfigDetails (Maybe Text)

The name of the AppImageConfig. Must be unique to your account.

appImageConfigDetails_lastModifiedTime :: Lens' AppImageConfigDetails (Maybe UTCTime)

When the AppImageConfig was last modified.