Network.AWS.SageMaker.Types.AppSpecification

Description

Documentation

data AppSpecification

Configuration to run a processing job in a specified container image.

See: newAppSpecification smart constructor.

Constructors

AppSpecification'

Fields

containerArguments :: Maybe (NonEmpty Text)

The arguments for a container used to run a processing job.

containerEntrypoint :: Maybe (NonEmpty Text)

The entrypoint for a container used to run a processing job.

imageUri :: Text

The container image to be run by the processing job.

newAppSpecification

Arguments

:: Text
$sel:imageUri:AppSpecification'

-> AppSpecification

Create a value of AppSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerArguments:AppSpecification', appSpecification_containerArguments - The arguments for a container used to run a processing job.

$sel:containerEntrypoint:AppSpecification', appSpecification_containerEntrypoint - The entrypoint for a container used to run a processing job.

$sel:imageUri:AppSpecification', appSpecification_imageUri - The container image to be run by the processing job.

appSpecification_containerArguments :: Lens' AppSpecification (Maybe (NonEmpty Text))

The arguments for a container used to run a processing job.

appSpecification_containerEntrypoint :: Lens' AppSpecification (Maybe (NonEmpty Text))

The entrypoint for a container used to run a processing job.

appSpecification_imageUri :: Lens' AppSpecification Text

The container image to be run by the processing job.