Network.AWS.SageMaker.Types.AsyncInferenceConfig

Description

Documentation

data AsyncInferenceConfig

Specifies configuration for how an endpoint performs asynchronous inference.

See: newAsyncInferenceConfig smart constructor.

Constructors

AsyncInferenceConfig'

Fields

clientConfig :: Maybe AsyncInferenceClientConfig

Configures the behavior of the client used by Amazon SageMaker to interact with the model container during asynchronous inference.

outputConfig :: AsyncInferenceOutputConfig

Specifies the configuration for asynchronous inference invocation outputs.

newAsyncInferenceConfig

Arguments

:: AsyncInferenceOutputConfig
$sel:outputConfig:AsyncInferenceConfig'

-> AsyncInferenceConfig

Create a value of AsyncInferenceConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientConfig:AsyncInferenceConfig', asyncInferenceConfig_clientConfig - Configures the behavior of the client used by Amazon SageMaker to interact with the model container during asynchronous inference.

$sel:outputConfig:AsyncInferenceConfig', asyncInferenceConfig_outputConfig - Specifies the configuration for asynchronous inference invocation outputs.

asyncInferenceConfig_clientConfig :: Lens' AsyncInferenceConfig (Maybe AsyncInferenceClientConfig)

Configures the behavior of the client used by Amazon SageMaker to interact with the model container during asynchronous inference.

asyncInferenceConfig_outputConfig :: Lens' AsyncInferenceConfig AsyncInferenceOutputConfig

Specifies the configuration for asynchronous inference invocation outputs.