Network.AWS.SageMaker.Types.AthenaDatasetDefinition

Description

Documentation

data AthenaDatasetDefinition

Configuration for Athena Dataset Definition input.

See: newAthenaDatasetDefinition smart constructor.

Constructors

AthenaDatasetDefinition'

Fields

outputCompression :: Maybe AthenaResultCompressionType

kmsKeyId :: Maybe Text

The Amazon Web Services Key Management Service (Amazon Web Services KMS) key that Amazon SageMaker uses to encrypt data generated from an Athena query execution.

workGroup :: Maybe Text

catalog :: Text

database :: Text

queryString :: Text

outputS3Uri :: Text

The location in Amazon S3 where Athena query results are stored.

outputFormat :: AthenaResultFormat

newAthenaDatasetDefinition

Arguments

:: Text
$sel:catalog:AthenaDatasetDefinition'

-> Text
$sel:database:AthenaDatasetDefinition'

-> Text
$sel:queryString:AthenaDatasetDefinition'

-> Text
$sel:outputS3Uri:AthenaDatasetDefinition'

-> AthenaResultFormat
$sel:outputFormat:AthenaDatasetDefinition'

-> AthenaDatasetDefinition

Create a value of AthenaDatasetDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outputCompression:AthenaDatasetDefinition', athenaDatasetDefinition_outputCompression - Undocumented member.

$sel:kmsKeyId:AthenaDatasetDefinition', athenaDatasetDefinition_kmsKeyId - The Amazon Web Services Key Management Service (Amazon Web Services KMS) key that Amazon SageMaker uses to encrypt data generated from an Athena query execution.

$sel:workGroup:AthenaDatasetDefinition', athenaDatasetDefinition_workGroup - Undocumented member.

$sel:catalog:AthenaDatasetDefinition', athenaDatasetDefinition_catalog - Undocumented member.

$sel:database:AthenaDatasetDefinition', athenaDatasetDefinition_database - Undocumented member.

$sel:queryString:AthenaDatasetDefinition', athenaDatasetDefinition_queryString - Undocumented member.

$sel:outputS3Uri:AthenaDatasetDefinition', athenaDatasetDefinition_outputS3Uri - The location in Amazon S3 where Athena query results are stored.

$sel:outputFormat:AthenaDatasetDefinition', athenaDatasetDefinition_outputFormat - Undocumented member.

athenaDatasetDefinition_outputCompression :: Lens' AthenaDatasetDefinition (Maybe AthenaResultCompressionType)

Undocumented member.

athenaDatasetDefinition_kmsKeyId :: Lens' AthenaDatasetDefinition (Maybe Text)

The Amazon Web Services Key Management Service (Amazon Web Services KMS) key that Amazon SageMaker uses to encrypt data generated from an Athena query execution.

athenaDatasetDefinition_workGroup :: Lens' AthenaDatasetDefinition (Maybe Text)

Undocumented member.

athenaDatasetDefinition_catalog :: Lens' AthenaDatasetDefinition Text

Undocumented member.

athenaDatasetDefinition_database :: Lens' AthenaDatasetDefinition Text

Undocumented member.

athenaDatasetDefinition_queryString :: Lens' AthenaDatasetDefinition Text

Undocumented member.

athenaDatasetDefinition_outputS3Uri :: Lens' AthenaDatasetDefinition Text

The location in Amazon S3 where Athena query results are stored.

athenaDatasetDefinition_outputFormat :: Lens' AthenaDatasetDefinition AthenaResultFormat

Undocumented member.