Network.AWS.SageMaker.Types.AutoMLCandidate

Description

Documentation

data AutoMLCandidate

Information about a candidate produced by an AutoML training job, including its status, steps, and other properties.

See: newAutoMLCandidate smart constructor.

Constructors

AutoMLCandidate'

Fields

candidateProperties :: Maybe CandidateProperties

The properties of an AutoML candidate job.

endTime :: Maybe POSIX

The end time.

inferenceContainers :: Maybe [AutoMLContainerDefinition]

Information about the inference container definitions.

failureReason :: Maybe Text

The failure reason.

finalAutoMLJobObjectiveMetric :: Maybe FinalAutoMLJobObjectiveMetric

candidateName :: Text

The name of the candidate.

objectiveStatus :: ObjectiveStatus

The objective's status.

candidateSteps :: [AutoMLCandidateStep]

Information about the candidate's steps.

candidateStatus :: CandidateStatus

The candidate's status.

creationTime :: POSIX

The creation time.

lastModifiedTime :: POSIX

The last modified time.

newAutoMLCandidate

Arguments

:: Text
$sel:candidateName:AutoMLCandidate'

-> ObjectiveStatus
$sel:objectiveStatus:AutoMLCandidate'

-> CandidateStatus
$sel:candidateStatus:AutoMLCandidate'

-> UTCTime
$sel:creationTime:AutoMLCandidate'

-> UTCTime
$sel:lastModifiedTime:AutoMLCandidate'

-> AutoMLCandidate

Create a value of AutoMLCandidate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:candidateProperties:AutoMLCandidate', autoMLCandidate_candidateProperties - The properties of an AutoML candidate job.

$sel:endTime:AutoMLCandidate', autoMLCandidate_endTime - The end time.

$sel:inferenceContainers:AutoMLCandidate', autoMLCandidate_inferenceContainers - Information about the inference container definitions.

$sel:failureReason:AutoMLCandidate', autoMLCandidate_failureReason - The failure reason.

$sel:finalAutoMLJobObjectiveMetric:AutoMLCandidate', autoMLCandidate_finalAutoMLJobObjectiveMetric - Undocumented member.

$sel:candidateName:AutoMLCandidate', autoMLCandidate_candidateName - The name of the candidate.

$sel:objectiveStatus:AutoMLCandidate', autoMLCandidate_objectiveStatus - The objective's status.

$sel:candidateSteps:AutoMLCandidate', autoMLCandidate_candidateSteps - Information about the candidate's steps.

$sel:candidateStatus:AutoMLCandidate', autoMLCandidate_candidateStatus - The candidate's status.

$sel:creationTime:AutoMLCandidate', autoMLCandidate_creationTime - The creation time.

$sel:lastModifiedTime:AutoMLCandidate', autoMLCandidate_lastModifiedTime - The last modified time.

autoMLCandidate_candidateProperties :: Lens' AutoMLCandidate (Maybe CandidateProperties)

The properties of an AutoML candidate job.

autoMLCandidate_endTime :: Lens' AutoMLCandidate (Maybe UTCTime)

The end time.

autoMLCandidate_inferenceContainers :: Lens' AutoMLCandidate (Maybe [AutoMLContainerDefinition])

Information about the inference container definitions.

autoMLCandidate_failureReason :: Lens' AutoMLCandidate (Maybe Text)

The failure reason.

autoMLCandidate_finalAutoMLJobObjectiveMetric :: Lens' AutoMLCandidate (Maybe FinalAutoMLJobObjectiveMetric)

Undocumented member.

autoMLCandidate_candidateName :: Lens' AutoMLCandidate Text

The name of the candidate.

autoMLCandidate_objectiveStatus :: Lens' AutoMLCandidate ObjectiveStatus

The objective's status.

autoMLCandidate_candidateSteps :: Lens' AutoMLCandidate [AutoMLCandidateStep]

Information about the candidate's steps.

autoMLCandidate_candidateStatus :: Lens' AutoMLCandidate CandidateStatus

The candidate's status.

autoMLCandidate_creationTime :: Lens' AutoMLCandidate UTCTime

The creation time.

autoMLCandidate_lastModifiedTime :: Lens' AutoMLCandidate UTCTime

The last modified time.