Network.AWS.SageMaker.Types.AutoMLCandidateStep

Description

Documentation

data AutoMLCandidateStep

Information about the steps for a candidate and what step it is working on.

See: newAutoMLCandidateStep smart constructor.

Constructors

AutoMLCandidateStep'

Fields

candidateStepType :: CandidateStepType

Whether the candidate is at the transform, training, or processing step.

candidateStepArn :: Text

The ARN for the candidate's step.

candidateStepName :: Text

The name for the candidate's step.

newAutoMLCandidateStep

Arguments

:: CandidateStepType
$sel:candidateStepType:AutoMLCandidateStep'

-> Text
$sel:candidateStepArn:AutoMLCandidateStep'

-> Text
$sel:candidateStepName:AutoMLCandidateStep'

-> AutoMLCandidateStep

Create a value of AutoMLCandidateStep with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:candidateStepType:AutoMLCandidateStep', autoMLCandidateStep_candidateStepType - Whether the candidate is at the transform, training, or processing step.

$sel:candidateStepArn:AutoMLCandidateStep', autoMLCandidateStep_candidateStepArn - The ARN for the candidate's step.

$sel:candidateStepName:AutoMLCandidateStep', autoMLCandidateStep_candidateStepName - The name for the candidate's step.

autoMLCandidateStep_candidateStepType :: Lens' AutoMLCandidateStep CandidateStepType

Whether the candidate is at the transform, training, or processing step.

autoMLCandidateStep_candidateStepArn :: Lens' AutoMLCandidateStep Text

The ARN for the candidate's step.

autoMLCandidateStep_candidateStepName :: Lens' AutoMLCandidateStep Text

The name for the candidate's step.