Network.AWS.SageMaker.Types.AutoMLJobSummary

Description

Documentation

data AutoMLJobSummary

Provides a summary about an AutoML job.

See: newAutoMLJobSummary smart constructor.

Constructors

AutoMLJobSummary'

Fields

endTime :: Maybe POSIX

The end time of an AutoML job.

partialFailureReasons :: Maybe (NonEmpty AutoMLPartialFailureReason)

The list of reasons for partial failures within an AutoML job.

failureReason :: Maybe Text

The failure reason of an AutoML job.

autoMLJobName :: Text

The name of the AutoML job you are requesting.

autoMLJobArn :: Text

The ARN of the AutoML job.

autoMLJobStatus :: AutoMLJobStatus

The status of the AutoML job.

autoMLJobSecondaryStatus :: AutoMLJobSecondaryStatus

The secondary status of the AutoML job.

creationTime :: POSIX

When the AutoML job was created.

lastModifiedTime :: POSIX

When the AutoML job was last modified.

newAutoMLJobSummary

Arguments

:: Text
$sel:autoMLJobName:AutoMLJobSummary'

-> Text
$sel:autoMLJobArn:AutoMLJobSummary'

-> AutoMLJobStatus
$sel:autoMLJobStatus:AutoMLJobSummary'

-> AutoMLJobSecondaryStatus
$sel:autoMLJobSecondaryStatus:AutoMLJobSummary'

-> UTCTime
$sel:creationTime:AutoMLJobSummary'

-> UTCTime
$sel:lastModifiedTime:AutoMLJobSummary'

-> AutoMLJobSummary

Create a value of AutoMLJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endTime:AutoMLJobSummary', autoMLJobSummary_endTime - The end time of an AutoML job.

$sel:partialFailureReasons:AutoMLJobSummary', autoMLJobSummary_partialFailureReasons - The list of reasons for partial failures within an AutoML job.

$sel:failureReason:AutoMLJobSummary', autoMLJobSummary_failureReason - The failure reason of an AutoML job.

$sel:autoMLJobName:AutoMLJobSummary', autoMLJobSummary_autoMLJobName - The name of the AutoML job you are requesting.

$sel:autoMLJobArn:AutoMLJobSummary', autoMLJobSummary_autoMLJobArn - The ARN of the AutoML job.

$sel:autoMLJobStatus:AutoMLJobSummary', autoMLJobSummary_autoMLJobStatus - The status of the AutoML job.

$sel:autoMLJobSecondaryStatus:AutoMLJobSummary', autoMLJobSummary_autoMLJobSecondaryStatus - The secondary status of the AutoML job.

$sel:creationTime:AutoMLJobSummary', autoMLJobSummary_creationTime - When the AutoML job was created.

$sel:lastModifiedTime:AutoMLJobSummary', autoMLJobSummary_lastModifiedTime - When the AutoML job was last modified.

autoMLJobSummary_endTime :: Lens' AutoMLJobSummary (Maybe UTCTime)

The end time of an AutoML job.

autoMLJobSummary_partialFailureReasons :: Lens' AutoMLJobSummary (Maybe (NonEmpty AutoMLPartialFailureReason))

The list of reasons for partial failures within an AutoML job.

autoMLJobSummary_failureReason :: Lens' AutoMLJobSummary (Maybe Text)

The failure reason of an AutoML job.

autoMLJobSummary_autoMLJobName :: Lens' AutoMLJobSummary Text

The name of the AutoML job you are requesting.

autoMLJobSummary_autoMLJobArn :: Lens' AutoMLJobSummary Text

The ARN of the AutoML job.

autoMLJobSummary_autoMLJobStatus :: Lens' AutoMLJobSummary AutoMLJobStatus

The status of the AutoML job.

autoMLJobSummary_autoMLJobSecondaryStatus :: Lens' AutoMLJobSummary AutoMLJobSecondaryStatus

The secondary status of the AutoML job.

autoMLJobSummary_creationTime :: Lens' AutoMLJobSummary UTCTime

When the AutoML job was created.

autoMLJobSummary_lastModifiedTime :: Lens' AutoMLJobSummary UTCTime

When the AutoML job was last modified.