Network.AWS.SageMaker.Types.CompilationJobSummary

Description

Documentation

data CompilationJobSummary

A summary of a model compilation job.

See: newCompilationJobSummary smart constructor.

Constructors

CompilationJobSummary'

Fields

compilationTargetPlatformArch :: Maybe TargetPlatformArch

The type of architecture that the model will run on after the compilation job has completed.

compilationStartTime :: Maybe POSIX

The time when the model compilation job started.

compilationTargetPlatformOs :: Maybe TargetPlatformOs

The type of OS that the model will run on after the compilation job has completed.

compilationTargetPlatformAccelerator :: Maybe TargetPlatformAccelerator

The type of accelerator that the model will run on after the compilation job has completed.

lastModifiedTime :: Maybe POSIX

The time when the model compilation job was last modified.

compilationEndTime :: Maybe POSIX

The time when the model compilation job completed.

compilationTargetDevice :: Maybe TargetDevice

The type of device that the model will run on after the compilation job has completed.

compilationJobName :: Text

The name of the model compilation job that you want a summary for.

compilationJobArn :: Text

The Amazon Resource Name (ARN) of the model compilation job.

creationTime :: POSIX

The time when the model compilation job was created.

compilationJobStatus :: CompilationJobStatus

The status of the model compilation job.

newCompilationJobSummary

Arguments

:: Text
$sel:compilationJobName:CompilationJobSummary'

-> Text
$sel:compilationJobArn:CompilationJobSummary'

-> UTCTime
$sel:creationTime:CompilationJobSummary'

-> CompilationJobStatus
$sel:compilationJobStatus:CompilationJobSummary'

-> CompilationJobSummary

Create a value of CompilationJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:compilationTargetPlatformArch:CompilationJobSummary', compilationJobSummary_compilationTargetPlatformArch - The type of architecture that the model will run on after the compilation job has completed.

$sel:compilationStartTime:CompilationJobSummary', compilationJobSummary_compilationStartTime - The time when the model compilation job started.

$sel:compilationTargetPlatformOs:CompilationJobSummary', compilationJobSummary_compilationTargetPlatformOs - The type of OS that the model will run on after the compilation job has completed.

$sel:compilationTargetPlatformAccelerator:CompilationJobSummary', compilationJobSummary_compilationTargetPlatformAccelerator - The type of accelerator that the model will run on after the compilation job has completed.

$sel:lastModifiedTime:CompilationJobSummary', compilationJobSummary_lastModifiedTime - The time when the model compilation job was last modified.

$sel:compilationEndTime:CompilationJobSummary', compilationJobSummary_compilationEndTime - The time when the model compilation job completed.

$sel:compilationTargetDevice:CompilationJobSummary', compilationJobSummary_compilationTargetDevice - The type of device that the model will run on after the compilation job has completed.

$sel:compilationJobName:CompilationJobSummary', compilationJobSummary_compilationJobName - The name of the model compilation job that you want a summary for.

$sel:compilationJobArn:CompilationJobSummary', compilationJobSummary_compilationJobArn - The Amazon Resource Name (ARN) of the model compilation job.

$sel:creationTime:CompilationJobSummary', compilationJobSummary_creationTime - The time when the model compilation job was created.

$sel:compilationJobStatus:CompilationJobSummary', compilationJobSummary_compilationJobStatus - The status of the model compilation job.

compilationJobSummary_compilationTargetPlatformArch :: Lens' CompilationJobSummary (Maybe TargetPlatformArch)

The type of architecture that the model will run on after the compilation job has completed.

compilationJobSummary_compilationStartTime :: Lens' CompilationJobSummary (Maybe UTCTime)

The time when the model compilation job started.

compilationJobSummary_compilationTargetPlatformOs :: Lens' CompilationJobSummary (Maybe TargetPlatformOs)

The type of OS that the model will run on after the compilation job has completed.

compilationJobSummary_compilationTargetPlatformAccelerator :: Lens' CompilationJobSummary (Maybe TargetPlatformAccelerator)

The type of accelerator that the model will run on after the compilation job has completed.

compilationJobSummary_lastModifiedTime :: Lens' CompilationJobSummary (Maybe UTCTime)

The time when the model compilation job was last modified.

compilationJobSummary_compilationEndTime :: Lens' CompilationJobSummary (Maybe UTCTime)

The time when the model compilation job completed.

compilationJobSummary_compilationTargetDevice :: Lens' CompilationJobSummary (Maybe TargetDevice)

The type of device that the model will run on after the compilation job has completed.

compilationJobSummary_compilationJobName :: Lens' CompilationJobSummary Text

The name of the model compilation job that you want a summary for.

compilationJobSummary_compilationJobArn :: Lens' CompilationJobSummary Text

The Amazon Resource Name (ARN) of the model compilation job.

compilationJobSummary_creationTime :: Lens' CompilationJobSummary UTCTime

The time when the model compilation job was created.

compilationJobSummary_compilationJobStatus :: Lens' CompilationJobSummary CompilationJobStatus

The status of the model compilation job.