Network.AWS.SageMaker.Types.DatasetDefinition

Description

Documentation

data DatasetDefinition

Configuration for Dataset Definition inputs. The Dataset Definition input must specify exactly one of either AthenaDatasetDefinition or RedshiftDatasetDefinition types.

See: newDatasetDefinition smart constructor.

Constructors

DatasetDefinition'

Fields

redshiftDatasetDefinition :: Maybe RedshiftDatasetDefinition

localPath :: Maybe Text

The local path where you want Amazon SageMaker to download the Dataset Definition inputs to run a processing job. LocalPath is an absolute path to the input data. This is a required parameter when AppManaged is False (default).

athenaDatasetDefinition :: Maybe AthenaDatasetDefinition

inputMode :: Maybe InputMode

Whether to use File or Pipe input mode. In File (default) mode, Amazon SageMaker copies the data from the input source onto the local Amazon Elastic Block Store (Amazon EBS) volumes before starting your training algorithm. This is the most commonly used input mode. In Pipe mode, Amazon SageMaker streams input data from the source directly to your algorithm without using the EBS volume.

dataDistributionType :: Maybe DataDistributionType

Whether the generated dataset is FullyReplicated or ShardedByS3Key (default).

newDatasetDefinition :: DatasetDefinition

Create a value of DatasetDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:redshiftDatasetDefinition:DatasetDefinition', datasetDefinition_redshiftDatasetDefinition - Undocumented member.

$sel:localPath:DatasetDefinition', datasetDefinition_localPath - The local path where you want Amazon SageMaker to download the Dataset Definition inputs to run a processing job. LocalPath is an absolute path to the input data. This is a required parameter when AppManaged is False (default).

$sel:athenaDatasetDefinition:DatasetDefinition', datasetDefinition_athenaDatasetDefinition - Undocumented member.

$sel:inputMode:DatasetDefinition', datasetDefinition_inputMode - Whether to use File or Pipe input mode. In File (default) mode, Amazon SageMaker copies the data from the input source onto the local Amazon Elastic Block Store (Amazon EBS) volumes before starting your training algorithm. This is the most commonly used input mode. In Pipe mode, Amazon SageMaker streams input data from the source directly to your algorithm without using the EBS volume.

$sel:dataDistributionType:DatasetDefinition', datasetDefinition_dataDistributionType - Whether the generated dataset is FullyReplicated or ShardedByS3Key (default).

datasetDefinition_redshiftDatasetDefinition :: Lens' DatasetDefinition (Maybe RedshiftDatasetDefinition)

Undocumented member.

datasetDefinition_localPath :: Lens' DatasetDefinition (Maybe Text)

The local path where you want Amazon SageMaker to download the Dataset Definition inputs to run a processing job. LocalPath is an absolute path to the input data. This is a required parameter when AppManaged is False (default).

datasetDefinition_athenaDatasetDefinition :: Lens' DatasetDefinition (Maybe AthenaDatasetDefinition)

Undocumented member.

datasetDefinition_inputMode :: Lens' DatasetDefinition (Maybe InputMode)

Whether to use File or Pipe input mode. In File (default) mode, Amazon SageMaker copies the data from the input source onto the local Amazon Elastic Block Store (Amazon EBS) volumes before starting your training algorithm. This is the most commonly used input mode. In Pipe mode, Amazon SageMaker streams input data from the source directly to your algorithm without using the EBS volume.

datasetDefinition_dataDistributionType :: Lens' DatasetDefinition (Maybe DataDistributionType)

Whether the generated dataset is FullyReplicated or ShardedByS3Key (default).