Network.AWS.SageMaker.Types.DeployedImage

Description

Documentation

data DeployedImage

Gets the Amazon EC2 Container Registry path of the docker image of the model that is hosted in this ProductionVariant.

If you used the registry/repository[:tag] form to specify the image path of the primary container when you created the model hosted in this ProductionVariant, the path resolves to a path of the form registry/repository[@digest]. A digest is a hash value that identifies a specific version of an image. For information about Amazon ECR paths, see Pulling an Image in the Amazon ECR User Guide.

See: newDeployedImage smart constructor.

Constructors

DeployedImage'

Fields

specifiedImage :: Maybe Text

The image path you specified when you created the model.

resolvedImage :: Maybe Text

The specific digest path of the image hosted in this ProductionVariant.

resolutionTime :: Maybe POSIX

The date and time when the image path for the model resolved to the ResolvedImage

newDeployedImage :: DeployedImage

Create a value of DeployedImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:specifiedImage:DeployedImage', deployedImage_specifiedImage - The image path you specified when you created the model.

$sel:resolvedImage:DeployedImage', deployedImage_resolvedImage - The specific digest path of the image hosted in this ProductionVariant.

$sel:resolutionTime:DeployedImage', deployedImage_resolutionTime - The date and time when the image path for the model resolved to the ResolvedImage

deployedImage_specifiedImage :: Lens' DeployedImage (Maybe Text)

The image path you specified when you created the model.

deployedImage_resolvedImage :: Lens' DeployedImage (Maybe Text)

The specific digest path of the image hosted in this ProductionVariant.

deployedImage_resolutionTime :: Lens' DeployedImage (Maybe UTCTime)

The date and time when the image path for the model resolved to the ResolvedImage