Network.AWS.SageMaker.Types.EdgeModelStat

Description

Documentation

data EdgeModelStat

Status of edge devices with this model.

See: newEdgeModelStat smart constructor.

Constructors

EdgeModelStat'

Fields

modelName :: Text

The name of the model.

modelVersion :: Text

The model version.

offlineDeviceCount :: Integer

The number of devices that have this model version and do not have a heart beat.

connectedDeviceCount :: Integer

The number of devices that have this model version and have a heart beat.

activeDeviceCount :: Integer

The number of devices that have this model version, a heart beat, and are currently running.

samplingDeviceCount :: Integer

The number of devices with this model version and are producing sample data.

newEdgeModelStat

Arguments

:: Text
$sel:modelName:EdgeModelStat'

-> Text
$sel:modelVersion:EdgeModelStat'

-> Integer
$sel:offlineDeviceCount:EdgeModelStat'

-> Integer
$sel:connectedDeviceCount:EdgeModelStat'

-> Integer
$sel:activeDeviceCount:EdgeModelStat'

-> Integer
$sel:samplingDeviceCount:EdgeModelStat'

-> EdgeModelStat

Create a value of EdgeModelStat with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelName:EdgeModelStat', edgeModelStat_modelName - The name of the model.

$sel:modelVersion:EdgeModelStat', edgeModelStat_modelVersion - The model version.

$sel:offlineDeviceCount:EdgeModelStat', edgeModelStat_offlineDeviceCount - The number of devices that have this model version and do not have a heart beat.

$sel:connectedDeviceCount:EdgeModelStat', edgeModelStat_connectedDeviceCount - The number of devices that have this model version and have a heart beat.

$sel:activeDeviceCount:EdgeModelStat', edgeModelStat_activeDeviceCount - The number of devices that have this model version, a heart beat, and are currently running.

$sel:samplingDeviceCount:EdgeModelStat', edgeModelStat_samplingDeviceCount - The number of devices with this model version and are producing sample data.

edgeModelStat_modelName :: Lens' EdgeModelStat Text

The name of the model.

edgeModelStat_modelVersion :: Lens' EdgeModelStat Text

The model version.

edgeModelStat_offlineDeviceCount :: Lens' EdgeModelStat Integer

The number of devices that have this model version and do not have a heart beat.

edgeModelStat_connectedDeviceCount :: Lens' EdgeModelStat Integer

The number of devices that have this model version and have a heart beat.

edgeModelStat_activeDeviceCount :: Lens' EdgeModelStat Integer

The number of devices that have this model version, a heart beat, and are currently running.

edgeModelStat_samplingDeviceCount :: Lens' EdgeModelStat Integer

The number of devices with this model version and are producing sample data.