Network.AWS.SageMaker.Types.EdgeOutputConfig

Description

Documentation

data EdgeOutputConfig

The output configuration.

See: newEdgeOutputConfig smart constructor.

Constructors

EdgeOutputConfig'

Fields

presetDeploymentConfig :: Maybe Text

The configuration used to create deployment artifacts. Specify configuration options with a JSON string. The available configuration options for each type are:

ComponentName (optional) - Name of the GreenGrass V2 component. If not specified, the default name generated consists of "SagemakerEdgeManager" and the name of your SageMaker Edge Manager packaging job.

ComponentDescription (optional) - Description of the component.

ComponentVersion (optional) - The version of the component.

Amazon Web Services IoT Greengrass uses semantic versions for components. Semantic versions follow a major.minor.patch number system. For example, version 1.0.0 represents the first major release for a component. For more information, see the semantic version specification.

PlatformOS (optional) - The name of the operating system for the platform. Supported platforms include Windows and Linux.

PlatformArchitecture (optional) - The processor architecture for the platform.

Supported architectures Windows include: Windows32_x86, Windows64_x64.

Supported architectures for Linux include: Linux x86_64, Linux ARMV8.

kmsKeyId :: Maybe Text

The Amazon Web Services Key Management Service (Amazon Web Services KMS) key that Amazon SageMaker uses to encrypt data on the storage volume after compilation job. If you don't provide a KMS key ID, Amazon SageMaker uses the default KMS key for Amazon S3 for your role's account.

presetDeploymentType :: Maybe EdgePresetDeploymentType

The deployment type SageMaker Edge Manager will create. Currently only supports Amazon Web Services IoT Greengrass Version 2 components.

s3OutputLocation :: Text

The Amazon Simple Storage (S3) bucker URI.

newEdgeOutputConfig

Arguments

:: Text
$sel:s3OutputLocation:EdgeOutputConfig'

-> EdgeOutputConfig

Create a value of EdgeOutputConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:presetDeploymentConfig:EdgeOutputConfig', edgeOutputConfig_presetDeploymentConfig - The configuration used to create deployment artifacts. Specify configuration options with a JSON string. The available configuration options for each type are:

ComponentName (optional) - Name of the GreenGrass V2 component. If not specified, the default name generated consists of "SagemakerEdgeManager" and the name of your SageMaker Edge Manager packaging job.

ComponentDescription (optional) - Description of the component.

ComponentVersion (optional) - The version of the component.

Amazon Web Services IoT Greengrass uses semantic versions for components. Semantic versions follow a major.minor.patch number system. For example, version 1.0.0 represents the first major release for a component. For more information, see the semantic version specification.

PlatformOS (optional) - The name of the operating system for the platform. Supported platforms include Windows and Linux.

PlatformArchitecture (optional) - The processor architecture for the platform.

Supported architectures Windows include: Windows32_x86, Windows64_x64.

Supported architectures for Linux include: Linux x86_64, Linux ARMV8.

$sel:kmsKeyId:EdgeOutputConfig', edgeOutputConfig_kmsKeyId - The Amazon Web Services Key Management Service (Amazon Web Services KMS) key that Amazon SageMaker uses to encrypt data on the storage volume after compilation job. If you don't provide a KMS key ID, Amazon SageMaker uses the default KMS key for Amazon S3 for your role's account.

$sel:presetDeploymentType:EdgeOutputConfig', edgeOutputConfig_presetDeploymentType - The deployment type SageMaker Edge Manager will create. Currently only supports Amazon Web Services IoT Greengrass Version 2 components.

$sel:s3OutputLocation:EdgeOutputConfig', edgeOutputConfig_s3OutputLocation - The Amazon Simple Storage (S3) bucker URI.

edgeOutputConfig_presetDeploymentConfig :: Lens' EdgeOutputConfig (Maybe Text)

The configuration used to create deployment artifacts. Specify configuration options with a JSON string. The available configuration options for each type are:

ComponentName (optional) - Name of the GreenGrass V2 component. If not specified, the default name generated consists of "SagemakerEdgeManager" and the name of your SageMaker Edge Manager packaging job.

ComponentDescription (optional) - Description of the component.

ComponentVersion (optional) - The version of the component.

Amazon Web Services IoT Greengrass uses semantic versions for components. Semantic versions follow a major.minor.patch number system. For example, version 1.0.0 represents the first major release for a component. For more information, see the semantic version specification.

PlatformOS (optional) - The name of the operating system for the platform. Supported platforms include Windows and Linux.

PlatformArchitecture (optional) - The processor architecture for the platform.

Supported architectures Windows include: Windows32_x86, Windows64_x64.

Supported architectures for Linux include: Linux x86_64, Linux ARMV8.

edgeOutputConfig_kmsKeyId :: Lens' EdgeOutputConfig (Maybe Text)

The Amazon Web Services Key Management Service (Amazon Web Services KMS) key that Amazon SageMaker uses to encrypt data on the storage volume after compilation job. If you don't provide a KMS key ID, Amazon SageMaker uses the default KMS key for Amazon S3 for your role's account.

edgeOutputConfig_presetDeploymentType :: Lens' EdgeOutputConfig (Maybe EdgePresetDeploymentType)

The deployment type SageMaker Edge Manager will create. Currently only supports Amazon Web Services IoT Greengrass Version 2 components.

edgeOutputConfig_s3OutputLocation :: Lens' EdgeOutputConfig Text

The Amazon Simple Storage (S3) bucker URI.