Network.AWS.SageMaker.Types.Endpoint

Description

Documentation

data Endpoint

A hosted endpoint for real-time inference.

See: newEndpoint smart constructor.

Constructors

Endpoint'

Fields

productionVariants :: Maybe (NonEmpty ProductionVariantSummary)

A list of the production variants hosted on the endpoint. Each production variant is a model.

monitoringSchedules :: Maybe [MonitoringSchedule]

A list of monitoring schedules for the endpoint. For information about model monitoring, see Amazon SageMaker Model Monitor.

failureReason :: Maybe Text

If the endpoint failed, the reason it failed.

tags :: Maybe [Tag]

A list of the tags associated with the endpoint. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

dataCaptureConfig :: Maybe DataCaptureConfigSummary

endpointName :: Text

The name of the endpoint.

endpointArn :: Text

The Amazon Resource Name (ARN) of the endpoint.

endpointConfigName :: Text

The endpoint configuration associated with the endpoint.

endpointStatus :: EndpointStatus

The status of the endpoint.

creationTime :: POSIX

The time that the endpoint was created.

lastModifiedTime :: POSIX

The last time the endpoint was modified.

newEndpoint

Arguments

:: Text
$sel:endpointName:Endpoint'

-> Text
$sel:endpointArn:Endpoint'

-> Text
$sel:endpointConfigName:Endpoint'

-> EndpointStatus
$sel:endpointStatus:Endpoint'

-> UTCTime
$sel:creationTime:Endpoint'

-> UTCTime
$sel:lastModifiedTime:Endpoint'

-> Endpoint

Create a value of Endpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:productionVariants:Endpoint', endpoint_productionVariants - A list of the production variants hosted on the endpoint. Each production variant is a model.

$sel:monitoringSchedules:Endpoint', endpoint_monitoringSchedules - A list of monitoring schedules for the endpoint. For information about model monitoring, see Amazon SageMaker Model Monitor.

$sel:failureReason:Endpoint', endpoint_failureReason - If the endpoint failed, the reason it failed.

$sel:tags:Endpoint', endpoint_tags - A list of the tags associated with the endpoint. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

$sel:dataCaptureConfig:Endpoint', endpoint_dataCaptureConfig - Undocumented member.

$sel:endpointName:Endpoint', endpoint_endpointName - The name of the endpoint.

$sel:endpointArn:Endpoint', endpoint_endpointArn - The Amazon Resource Name (ARN) of the endpoint.

$sel:endpointConfigName:Endpoint', endpoint_endpointConfigName - The endpoint configuration associated with the endpoint.

$sel:endpointStatus:Endpoint', endpoint_endpointStatus - The status of the endpoint.

$sel:creationTime:Endpoint', endpoint_creationTime - The time that the endpoint was created.

$sel:lastModifiedTime:Endpoint', endpoint_lastModifiedTime - The last time the endpoint was modified.

endpoint_productionVariants :: Lens' Endpoint (Maybe (NonEmpty ProductionVariantSummary))

A list of the production variants hosted on the endpoint. Each production variant is a model.

endpoint_monitoringSchedules :: Lens' Endpoint (Maybe [MonitoringSchedule])

A list of monitoring schedules for the endpoint. For information about model monitoring, see Amazon SageMaker Model Monitor.

endpoint_failureReason :: Lens' Endpoint (Maybe Text)

If the endpoint failed, the reason it failed.

endpoint_tags :: Lens' Endpoint (Maybe [Tag])

A list of the tags associated with the endpoint. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

endpoint_dataCaptureConfig :: Lens' Endpoint (Maybe DataCaptureConfigSummary)

Undocumented member.

endpoint_endpointName :: Lens' Endpoint Text

The name of the endpoint.

endpoint_endpointArn :: Lens' Endpoint Text

The Amazon Resource Name (ARN) of the endpoint.

endpoint_endpointConfigName :: Lens' Endpoint Text

The endpoint configuration associated with the endpoint.

endpoint_endpointStatus :: Lens' Endpoint EndpointStatus

The status of the endpoint.

endpoint_creationTime :: Lens' Endpoint UTCTime

The time that the endpoint was created.

endpoint_lastModifiedTime :: Lens' Endpoint UTCTime

The last time the endpoint was modified.