Network.AWS.SageMaker.Types.EndpointInput

Description

Documentation

data EndpointInput

Input object for the endpoint

See: newEndpointInput smart constructor.

Constructors

EndpointInput'

Fields

endTimeOffset :: Maybe Text

If specified, monitoring jobs substract this time from the end time. For information about using offsets for scheduling monitoring jobs, see Schedule Model Quality Monitoring Jobs.

inferenceAttribute :: Maybe Text

The attribute of the input data that represents the ground truth label.

s3InputMode :: Maybe ProcessingS3InputMode

Whether the Pipe or File is used as the input mode for transfering data for the monitoring job. Pipe mode is recommended for large datasets. File mode is useful for small files that fit in memory. Defaults to File.

s3DataDistributionType :: Maybe ProcessingS3DataDistributionType

Whether input data distributed in Amazon S3 is fully replicated or sharded by an S3 key. Defaults to FullyReplicated

probabilityAttribute :: Maybe Text

In a classification problem, the attribute that represents the class probability.

probabilityThresholdAttribute :: Maybe Double

The threshold for the class probability to be evaluated as a positive result.

featuresAttribute :: Maybe Text

The attributes of the input data that are the input features.

startTimeOffset :: Maybe Text

If specified, monitoring jobs substract this time from the start time. For information about using offsets for scheduling monitoring jobs, see Schedule Model Quality Monitoring Jobs.

endpointName :: Text

An endpoint in customer's account which has enabled DataCaptureConfig enabled.

localPath :: Text

Path to the filesystem where the endpoint data is available to the container.

newEndpointInput

Arguments

:: Text
$sel:endpointName:EndpointInput'

-> Text
$sel:localPath:EndpointInput'

-> EndpointInput

Create a value of EndpointInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endTimeOffset:EndpointInput', endpointInput_endTimeOffset - If specified, monitoring jobs substract this time from the end time. For information about using offsets for scheduling monitoring jobs, see Schedule Model Quality Monitoring Jobs.

$sel:inferenceAttribute:EndpointInput', endpointInput_inferenceAttribute - The attribute of the input data that represents the ground truth label.

$sel:s3InputMode:EndpointInput', endpointInput_s3InputMode - Whether the Pipe or File is used as the input mode for transfering data for the monitoring job. Pipe mode is recommended for large datasets. File mode is useful for small files that fit in memory. Defaults to File.

$sel:s3DataDistributionType:EndpointInput', endpointInput_s3DataDistributionType - Whether input data distributed in Amazon S3 is fully replicated or sharded by an S3 key. Defaults to FullyReplicated

$sel:probabilityAttribute:EndpointInput', endpointInput_probabilityAttribute - In a classification problem, the attribute that represents the class probability.

$sel:probabilityThresholdAttribute:EndpointInput', endpointInput_probabilityThresholdAttribute - The threshold for the class probability to be evaluated as a positive result.

$sel:featuresAttribute:EndpointInput', endpointInput_featuresAttribute - The attributes of the input data that are the input features.

$sel:startTimeOffset:EndpointInput', endpointInput_startTimeOffset - If specified, monitoring jobs substract this time from the start time. For information about using offsets for scheduling monitoring jobs, see Schedule Model Quality Monitoring Jobs.

$sel:endpointName:EndpointInput', endpointInput_endpointName - An endpoint in customer's account which has enabled DataCaptureConfig enabled.

$sel:localPath:EndpointInput', endpointInput_localPath - Path to the filesystem where the endpoint data is available to the container.

endpointInput_endTimeOffset :: Lens' EndpointInput (Maybe Text)

If specified, monitoring jobs substract this time from the end time. For information about using offsets for scheduling monitoring jobs, see Schedule Model Quality Monitoring Jobs.

endpointInput_inferenceAttribute :: Lens' EndpointInput (Maybe Text)

The attribute of the input data that represents the ground truth label.

endpointInput_s3InputMode :: Lens' EndpointInput (Maybe ProcessingS3InputMode)

Whether the Pipe or File is used as the input mode for transfering data for the monitoring job. Pipe mode is recommended for large datasets. File mode is useful for small files that fit in memory. Defaults to File.

endpointInput_s3DataDistributionType :: Lens' EndpointInput (Maybe ProcessingS3DataDistributionType)

Whether input data distributed in Amazon S3 is fully replicated or sharded by an S3 key. Defaults to FullyReplicated

endpointInput_probabilityAttribute :: Lens' EndpointInput (Maybe Text)

In a classification problem, the attribute that represents the class probability.

endpointInput_probabilityThresholdAttribute :: Lens' EndpointInput (Maybe Double)

The threshold for the class probability to be evaluated as a positive result.

endpointInput_featuresAttribute :: Lens' EndpointInput (Maybe Text)

The attributes of the input data that are the input features.

endpointInput_startTimeOffset :: Lens' EndpointInput (Maybe Text)

If specified, monitoring jobs substract this time from the start time. For information about using offsets for scheduling monitoring jobs, see Schedule Model Quality Monitoring Jobs.

endpointInput_endpointName :: Lens' EndpointInput Text

An endpoint in customer's account which has enabled DataCaptureConfig enabled.

endpointInput_localPath :: Lens' EndpointInput Text

Path to the filesystem where the endpoint data is available to the container.