Network.AWS.SageMaker.Types.Experiment

Description

Documentation

data Experiment

The properties of an experiment as returned by the Search API.

See: newExperiment smart constructor.

Constructors

Experiment'

Fields

creationTime :: Maybe POSIX

When the experiment was created.

experimentArn :: Maybe Text

The Amazon Resource Name (ARN) of the experiment.

source :: Maybe ExperimentSource

tags :: Maybe [Tag]

The list of tags that are associated with the experiment. You can use Search API to search on the tags.

lastModifiedTime :: Maybe POSIX

When the experiment was last modified.

description :: Maybe Text

The description of the experiment.

experimentName :: Maybe Text

The name of the experiment.

createdBy :: Maybe UserContext

Who created the experiment.

lastModifiedBy :: Maybe UserContext

displayName :: Maybe Text

The name of the experiment as displayed. If DisplayName isn't specified, ExperimentName is displayed.

newExperiment :: Experiment

Create a value of Experiment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Experiment', experiment_creationTime - When the experiment was created.

$sel:experimentArn:Experiment', experiment_experimentArn - The Amazon Resource Name (ARN) of the experiment.

$sel:source:Experiment', experiment_source - Undocumented member.

$sel:tags:Experiment', experiment_tags - The list of tags that are associated with the experiment. You can use Search API to search on the tags.

$sel:lastModifiedTime:Experiment', experiment_lastModifiedTime - When the experiment was last modified.

$sel:description:Experiment', experiment_description - The description of the experiment.

$sel:experimentName:Experiment', experiment_experimentName - The name of the experiment.

$sel:createdBy:Experiment', experiment_createdBy - Who created the experiment.

$sel:lastModifiedBy:Experiment', experiment_lastModifiedBy - Undocumented member.

$sel:displayName:Experiment', experiment_displayName - The name of the experiment as displayed. If DisplayName isn't specified, ExperimentName is displayed.

experiment_creationTime :: Lens' Experiment (Maybe UTCTime)

When the experiment was created.

experiment_experimentArn :: Lens' Experiment (Maybe Text)

The Amazon Resource Name (ARN) of the experiment.

experiment_source :: Lens' Experiment (Maybe ExperimentSource)

Undocumented member.

experiment_tags :: Lens' Experiment (Maybe [Tag])

The list of tags that are associated with the experiment. You can use Search API to search on the tags.

experiment_lastModifiedTime :: Lens' Experiment (Maybe UTCTime)

When the experiment was last modified.

experiment_description :: Lens' Experiment (Maybe Text)

The description of the experiment.

experiment_experimentName :: Lens' Experiment (Maybe Text)

The name of the experiment.

experiment_createdBy :: Lens' Experiment (Maybe UserContext)

Who created the experiment.

experiment_lastModifiedBy :: Lens' Experiment (Maybe UserContext)

Undocumented member.

experiment_displayName :: Lens' Experiment (Maybe Text)

The name of the experiment as displayed. If DisplayName isn't specified, ExperimentName is displayed.