Network.AWS.SageMaker.Types.FeatureGroupSummary

Description

Documentation

data FeatureGroupSummary

The name, Arn, CreationTime, FeatureGroup values, LastUpdatedTime and EnableOnlineStorage status of a FeatureGroup.

See: newFeatureGroupSummary smart constructor.

Constructors

FeatureGroupSummary'

Fields

featureGroupStatus :: Maybe FeatureGroupStatus

The status of a FeatureGroup. The status can be any of the following: Creating, Created, CreateFail, Deleting or DetailFail.

offlineStoreStatus :: Maybe OfflineStoreStatus

Notifies you if replicating data into the OfflineStore has failed. Returns either: Active or Blocked.

featureGroupName :: Text

The name of FeatureGroup.

featureGroupArn :: Text

Unique identifier for the FeatureGroup.

creationTime :: POSIX

A timestamp indicating the time of creation time of the FeatureGroup.

newFeatureGroupSummary

Arguments

:: Text
$sel:featureGroupName:FeatureGroupSummary'

-> Text
$sel:featureGroupArn:FeatureGroupSummary'

-> UTCTime
$sel:creationTime:FeatureGroupSummary'

-> FeatureGroupSummary

Create a value of FeatureGroupSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureGroupStatus:FeatureGroupSummary', featureGroupSummary_featureGroupStatus - The status of a FeatureGroup. The status can be any of the following: Creating, Created, CreateFail, Deleting or DetailFail.

$sel:offlineStoreStatus:FeatureGroupSummary', featureGroupSummary_offlineStoreStatus - Notifies you if replicating data into the OfflineStore has failed. Returns either: Active or Blocked.

$sel:featureGroupName:FeatureGroupSummary', featureGroupSummary_featureGroupName - The name of FeatureGroup.

$sel:featureGroupArn:FeatureGroupSummary', featureGroupSummary_featureGroupArn - Unique identifier for the FeatureGroup.

$sel:creationTime:FeatureGroupSummary', featureGroupSummary_creationTime - A timestamp indicating the time of creation time of the FeatureGroup.

featureGroupSummary_featureGroupStatus :: Lens' FeatureGroupSummary (Maybe FeatureGroupStatus)

The status of a FeatureGroup. The status can be any of the following: Creating, Created, CreateFail, Deleting or DetailFail.

featureGroupSummary_offlineStoreStatus :: Lens' FeatureGroupSummary (Maybe OfflineStoreStatus)

Notifies you if replicating data into the OfflineStore has failed. Returns either: Active or Blocked.

featureGroupSummary_featureGroupName :: Lens' FeatureGroupSummary Text

The name of FeatureGroup.

featureGroupSummary_featureGroupArn :: Lens' FeatureGroupSummary Text

Unique identifier for the FeatureGroup.

featureGroupSummary_creationTime :: Lens' FeatureGroupSummary UTCTime

A timestamp indicating the time of creation time of the FeatureGroup.