Network.AWS.SageMaker.Types.FlowDefinitionSummary

Description

Documentation

data FlowDefinitionSummary

Contains summary information about the flow definition.

See: newFlowDefinitionSummary smart constructor.

Constructors

FlowDefinitionSummary'

Fields

failureReason :: Maybe Text

The reason why the flow definition creation failed. A failure reason is returned only when the flow definition status is Failed.

flowDefinitionName :: Text

The name of the flow definition.

flowDefinitionArn :: Text

The Amazon Resource Name (ARN) of the flow definition.

flowDefinitionStatus :: FlowDefinitionStatus

The status of the flow definition. Valid values:

creationTime :: POSIX

The timestamp when SageMaker created the flow definition.

newFlowDefinitionSummary

Arguments

:: Text
$sel:flowDefinitionName:FlowDefinitionSummary'

-> Text
$sel:flowDefinitionArn:FlowDefinitionSummary'

-> FlowDefinitionStatus
$sel:flowDefinitionStatus:FlowDefinitionSummary'

-> UTCTime
$sel:creationTime:FlowDefinitionSummary'

-> FlowDefinitionSummary

Create a value of FlowDefinitionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:FlowDefinitionSummary', flowDefinitionSummary_failureReason - The reason why the flow definition creation failed. A failure reason is returned only when the flow definition status is Failed.

$sel:flowDefinitionName:FlowDefinitionSummary', flowDefinitionSummary_flowDefinitionName - The name of the flow definition.

$sel:flowDefinitionArn:FlowDefinitionSummary', flowDefinitionSummary_flowDefinitionArn - The Amazon Resource Name (ARN) of the flow definition.

$sel:flowDefinitionStatus:FlowDefinitionSummary', flowDefinitionSummary_flowDefinitionStatus - The status of the flow definition. Valid values:

$sel:creationTime:FlowDefinitionSummary', flowDefinitionSummary_creationTime - The timestamp when SageMaker created the flow definition.

flowDefinitionSummary_failureReason :: Lens' FlowDefinitionSummary (Maybe Text)

The reason why the flow definition creation failed. A failure reason is returned only when the flow definition status is Failed.

flowDefinitionSummary_flowDefinitionName :: Lens' FlowDefinitionSummary Text

The name of the flow definition.

flowDefinitionSummary_flowDefinitionArn :: Lens' FlowDefinitionSummary Text

The Amazon Resource Name (ARN) of the flow definition.

flowDefinitionSummary_flowDefinitionStatus :: Lens' FlowDefinitionSummary FlowDefinitionStatus

The status of the flow definition. Valid values:

flowDefinitionSummary_creationTime :: Lens' FlowDefinitionSummary UTCTime

The timestamp when SageMaker created the flow definition.