Network.AWS.SageMaker.Types.HumanLoopConfig

Description

Documentation

data HumanLoopConfig

Describes the work to be performed by human workers.

See: newHumanLoopConfig smart constructor.

Constructors

HumanLoopConfig'

Fields

taskKeywords :: Maybe (NonEmpty Text)

Keywords used to describe the task so that workers can discover the task.

taskTimeLimitInSeconds :: Maybe Natural

The amount of time that a worker has to complete a task. The default value is 3,600 seconds (1 hour).

taskAvailabilityLifetimeInSeconds :: Maybe Natural

The length of time that a task remains available for review by human workers.

publicWorkforceTaskPrice :: Maybe PublicWorkforceTaskPrice

workteamArn :: Text

Amazon Resource Name (ARN) of a team of workers. To learn more about the types of workforces and work teams you can create and use with Amazon A2I, see Create and Manage Workforces.

humanTaskUiArn :: Text

The Amazon Resource Name (ARN) of the human task user interface.

You can use standard HTML and Crowd HTML Elements to create a custom worker task template. You use this template to create a human task UI.

To learn how to create a custom HTML template, see Create Custom Worker Task Template.

To learn how to create a human task UI, which is a worker task template that can be used in a flow definition, see Create and Delete a Worker Task Templates.

taskTitle :: Text

A title for the human worker task.

taskDescription :: Text

A description for the human worker task.

taskCount :: Natural

The number of distinct workers who will perform the same task on each object. For example, if TaskCount is set to 3 for an image classification labeling job, three workers will classify each input image. Increasing TaskCount can improve label accuracy.

newHumanLoopConfig

Arguments

:: Text
$sel:workteamArn:HumanLoopConfig'

-> Text
$sel:humanTaskUiArn:HumanLoopConfig'

-> Text
$sel:taskTitle:HumanLoopConfig'

-> Text
$sel:taskDescription:HumanLoopConfig'

-> Natural
$sel:taskCount:HumanLoopConfig'

-> HumanLoopConfig

Create a value of HumanLoopConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskKeywords:HumanLoopConfig', humanLoopConfig_taskKeywords - Keywords used to describe the task so that workers can discover the task.

$sel:taskTimeLimitInSeconds:HumanLoopConfig', humanLoopConfig_taskTimeLimitInSeconds - The amount of time that a worker has to complete a task. The default value is 3,600 seconds (1 hour).

$sel:taskAvailabilityLifetimeInSeconds:HumanLoopConfig', humanLoopConfig_taskAvailabilityLifetimeInSeconds - The length of time that a task remains available for review by human workers.

$sel:publicWorkforceTaskPrice:HumanLoopConfig', humanLoopConfig_publicWorkforceTaskPrice - Undocumented member.

$sel:workteamArn:HumanLoopConfig', humanLoopConfig_workteamArn - Amazon Resource Name (ARN) of a team of workers. To learn more about the types of workforces and work teams you can create and use with Amazon A2I, see Create and Manage Workforces.

$sel:humanTaskUiArn:HumanLoopConfig', humanLoopConfig_humanTaskUiArn - The Amazon Resource Name (ARN) of the human task user interface.

You can use standard HTML and Crowd HTML Elements to create a custom worker task template. You use this template to create a human task UI.

To learn how to create a custom HTML template, see Create Custom Worker Task Template.

To learn how to create a human task UI, which is a worker task template that can be used in a flow definition, see Create and Delete a Worker Task Templates.

$sel:taskTitle:HumanLoopConfig', humanLoopConfig_taskTitle - A title for the human worker task.

$sel:taskDescription:HumanLoopConfig', humanLoopConfig_taskDescription - A description for the human worker task.

$sel:taskCount:HumanLoopConfig', humanLoopConfig_taskCount - The number of distinct workers who will perform the same task on each object. For example, if TaskCount is set to 3 for an image classification labeling job, three workers will classify each input image. Increasing TaskCount can improve label accuracy.

humanLoopConfig_taskKeywords :: Lens' HumanLoopConfig (Maybe (NonEmpty Text))

Keywords used to describe the task so that workers can discover the task.

humanLoopConfig_taskTimeLimitInSeconds :: Lens' HumanLoopConfig (Maybe Natural)

The amount of time that a worker has to complete a task. The default value is 3,600 seconds (1 hour).

humanLoopConfig_taskAvailabilityLifetimeInSeconds :: Lens' HumanLoopConfig (Maybe Natural)

The length of time that a task remains available for review by human workers.

humanLoopConfig_publicWorkforceTaskPrice :: Lens' HumanLoopConfig (Maybe PublicWorkforceTaskPrice)

Undocumented member.

humanLoopConfig_workteamArn :: Lens' HumanLoopConfig Text

Amazon Resource Name (ARN) of a team of workers. To learn more about the types of workforces and work teams you can create and use with Amazon A2I, see Create and Manage Workforces.

humanLoopConfig_humanTaskUiArn :: Lens' HumanLoopConfig Text

The Amazon Resource Name (ARN) of the human task user interface.

You can use standard HTML and Crowd HTML Elements to create a custom worker task template. You use this template to create a human task UI.

To learn how to create a custom HTML template, see Create Custom Worker Task Template.

To learn how to create a human task UI, which is a worker task template that can be used in a flow definition, see Create and Delete a Worker Task Templates.

humanLoopConfig_taskTitle :: Lens' HumanLoopConfig Text

A title for the human worker task.

humanLoopConfig_taskDescription :: Lens' HumanLoopConfig Text

A description for the human worker task.

humanLoopConfig_taskCount :: Lens' HumanLoopConfig Natural

The number of distinct workers who will perform the same task on each object. For example, if TaskCount is set to 3 for an image classification labeling job, three workers will classify each input image. Increasing TaskCount can improve label accuracy.