Network.AWS.SageMaker.Types.HumanLoopRequestSource

Description

Documentation

data HumanLoopRequestSource

Container for configuring the source of human task requests.

See: newHumanLoopRequestSource smart constructor.

Constructors

HumanLoopRequestSource'

Fields

awsManagedHumanLoopRequestSource :: AwsManagedHumanLoopRequestSource

Specifies whether Amazon Rekognition or Amazon Textract are used as the integration source. The default field settings and JSON parsing rules are different based on the integration source. Valid values:

newHumanLoopRequestSource

Arguments

:: AwsManagedHumanLoopRequestSource
$sel:awsManagedHumanLoopRequestSource:HumanLoopRequestSource'

-> HumanLoopRequestSource

Create a value of HumanLoopRequestSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsManagedHumanLoopRequestSource:HumanLoopRequestSource', humanLoopRequestSource_awsManagedHumanLoopRequestSource - Specifies whether Amazon Rekognition or Amazon Textract are used as the integration source. The default field settings and JSON parsing rules are different based on the integration source. Valid values:

humanLoopRequestSource_awsManagedHumanLoopRequestSource :: Lens' HumanLoopRequestSource AwsManagedHumanLoopRequestSource

Specifies whether Amazon Rekognition or Amazon Textract are used as the integration source. The default field settings and JSON parsing rules are different based on the integration source. Valid values: