Network.AWS.SageMaker.Types.HyperParameterAlgorithmSpecification

Description

Documentation

data HyperParameterAlgorithmSpecification

Specifies which training algorithm to use for training jobs that a hyperparameter tuning job launches and the metrics to monitor.

See: newHyperParameterAlgorithmSpecification smart constructor.

Constructors

HyperParameterAlgorithmSpecification'

Fields

trainingImage :: Maybe Text

The registry path of the Docker image that contains the training algorithm. For information about Docker registry paths for built-in algorithms, see Algorithms Provided by Amazon SageMaker: Common Parameters. Amazon SageMaker supports both registry/repository[:tag] and registry/repository[@digest] image path formats. For more information, see Using Your Own Algorithms with Amazon SageMaker.

metricDefinitions :: Maybe [MetricDefinition]

An array of MetricDefinition objects that specify the metrics that the algorithm emits.

algorithmName :: Maybe Text

The name of the resource algorithm to use for the hyperparameter tuning job. If you specify a value for this parameter, do not specify a value for TrainingImage.

trainingInputMode :: TrainingInputMode

The input mode that the algorithm supports: File or Pipe. In File input mode, Amazon SageMaker downloads the training data from Amazon S3 to the storage volume that is attached to the training instance and mounts the directory to the Docker volume for the training container. In Pipe input mode, Amazon SageMaker streams data directly from Amazon S3 to the container.

If you specify File mode, make sure that you provision the storage volume that is attached to the training instance with enough capacity to accommodate the training data downloaded from Amazon S3, the model artifacts, and intermediate information.

For more information about input modes, see Algorithms.

newHyperParameterAlgorithmSpecification

Arguments

:: TrainingInputMode
$sel:trainingInputMode:HyperParameterAlgorithmSpecification'

-> HyperParameterAlgorithmSpecification

Create a value of HyperParameterAlgorithmSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trainingImage:HyperParameterAlgorithmSpecification', hyperParameterAlgorithmSpecification_trainingImage - The registry path of the Docker image that contains the training algorithm. For information about Docker registry paths for built-in algorithms, see Algorithms Provided by Amazon SageMaker: Common Parameters. Amazon SageMaker supports both registry/repository[:tag] and registry/repository[@digest] image path formats. For more information, see Using Your Own Algorithms with Amazon SageMaker.

$sel:metricDefinitions:HyperParameterAlgorithmSpecification', hyperParameterAlgorithmSpecification_metricDefinitions - An array of MetricDefinition objects that specify the metrics that the algorithm emits.

$sel:algorithmName:HyperParameterAlgorithmSpecification', hyperParameterAlgorithmSpecification_algorithmName - The name of the resource algorithm to use for the hyperparameter tuning job. If you specify a value for this parameter, do not specify a value for TrainingImage.

$sel:trainingInputMode:HyperParameterAlgorithmSpecification', hyperParameterAlgorithmSpecification_trainingInputMode - The input mode that the algorithm supports: File or Pipe. In File input mode, Amazon SageMaker downloads the training data from Amazon S3 to the storage volume that is attached to the training instance and mounts the directory to the Docker volume for the training container. In Pipe input mode, Amazon SageMaker streams data directly from Amazon S3 to the container.

If you specify File mode, make sure that you provision the storage volume that is attached to the training instance with enough capacity to accommodate the training data downloaded from Amazon S3, the model artifacts, and intermediate information.

For more information about input modes, see Algorithms.

hyperParameterAlgorithmSpecification_trainingImage :: Lens' HyperParameterAlgorithmSpecification (Maybe Text)

The registry path of the Docker image that contains the training algorithm. For information about Docker registry paths for built-in algorithms, see Algorithms Provided by Amazon SageMaker: Common Parameters. Amazon SageMaker supports both registry/repository[:tag] and registry/repository[@digest] image path formats. For more information, see Using Your Own Algorithms with Amazon SageMaker.

hyperParameterAlgorithmSpecification_metricDefinitions :: Lens' HyperParameterAlgorithmSpecification (Maybe [MetricDefinition])

An array of MetricDefinition objects that specify the metrics that the algorithm emits.

hyperParameterAlgorithmSpecification_algorithmName :: Lens' HyperParameterAlgorithmSpecification (Maybe Text)

The name of the resource algorithm to use for the hyperparameter tuning job. If you specify a value for this parameter, do not specify a value for TrainingImage.

hyperParameterAlgorithmSpecification_trainingInputMode :: Lens' HyperParameterAlgorithmSpecification TrainingInputMode

The input mode that the algorithm supports: File or Pipe. In File input mode, Amazon SageMaker downloads the training data from Amazon S3 to the storage volume that is attached to the training instance and mounts the directory to the Docker volume for the training container. In Pipe input mode, Amazon SageMaker streams data directly from Amazon S3 to the container.

If you specify File mode, make sure that you provision the storage volume that is attached to the training instance with enough capacity to accommodate the training data downloaded from Amazon S3, the model artifacts, and intermediate information.

For more information about input modes, see Algorithms.