Network.AWS.SageMaker.Types.HyperParameterSpecification

Description

Documentation

data HyperParameterSpecification

Defines a hyperparameter to be used by an algorithm.

See: newHyperParameterSpecification smart constructor.

Constructors

HyperParameterSpecification'

Fields

isTunable :: Maybe Bool

Indicates whether this hyperparameter is tunable in a hyperparameter tuning job.

range :: Maybe ParameterRange

The allowed range for this hyperparameter.

defaultValue :: Maybe Text

The default value for this hyperparameter. If a default value is specified, a hyperparameter cannot be required.

isRequired :: Maybe Bool

Indicates whether this hyperparameter is required.

description :: Maybe Text

A brief description of the hyperparameter.

name :: Text

The name of this hyperparameter. The name must be unique.

type' :: ParameterType

The type of this hyperparameter. The valid types are Integer, Continuous, Categorical, and FreeText.

newHyperParameterSpecification

Arguments

:: Text
$sel:name:HyperParameterSpecification'

-> ParameterType
$sel:type':HyperParameterSpecification'

-> HyperParameterSpecification

Create a value of HyperParameterSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isTunable:HyperParameterSpecification', hyperParameterSpecification_isTunable - Indicates whether this hyperparameter is tunable in a hyperparameter tuning job.

$sel:range:HyperParameterSpecification', hyperParameterSpecification_range - The allowed range for this hyperparameter.

$sel:defaultValue:HyperParameterSpecification', hyperParameterSpecification_defaultValue - The default value for this hyperparameter. If a default value is specified, a hyperparameter cannot be required.

$sel:isRequired:HyperParameterSpecification', hyperParameterSpecification_isRequired - Indicates whether this hyperparameter is required.

$sel:description:HyperParameterSpecification', hyperParameterSpecification_description - A brief description of the hyperparameter.

$sel:name:HyperParameterSpecification', hyperParameterSpecification_name - The name of this hyperparameter. The name must be unique.

$sel:type':HyperParameterSpecification', hyperParameterSpecification_type - The type of this hyperparameter. The valid types are Integer, Continuous, Categorical, and FreeText.

hyperParameterSpecification_isTunable :: Lens' HyperParameterSpecification (Maybe Bool)

Indicates whether this hyperparameter is tunable in a hyperparameter tuning job.

hyperParameterSpecification_range :: Lens' HyperParameterSpecification (Maybe ParameterRange)

The allowed range for this hyperparameter.

hyperParameterSpecification_defaultValue :: Lens' HyperParameterSpecification (Maybe Text)

The default value for this hyperparameter. If a default value is specified, a hyperparameter cannot be required.

hyperParameterSpecification_isRequired :: Lens' HyperParameterSpecification (Maybe Bool)

Indicates whether this hyperparameter is required.

hyperParameterSpecification_description :: Lens' HyperParameterSpecification (Maybe Text)

A brief description of the hyperparameter.

hyperParameterSpecification_name :: Lens' HyperParameterSpecification Text

The name of this hyperparameter. The name must be unique.

hyperParameterSpecification_type :: Lens' HyperParameterSpecification ParameterType

The type of this hyperparameter. The valid types are Integer, Continuous, Categorical, and FreeText.