Network.AWS.SageMaker.Types.HyperParameterTrainingJobDefinition

Description

Documentation

data HyperParameterTrainingJobDefinition

Defines the training jobs launched by a hyperparameter tuning job.

See: newHyperParameterTrainingJobDefinition smart constructor.

Constructors

HyperParameterTrainingJobDefinition'

Fields

tuningObjective :: Maybe HyperParameterTuningJobObjective

checkpointConfig :: Maybe CheckpointConfig

hyperParameterRanges :: Maybe ParameterRanges

retryStrategy :: Maybe RetryStrategy

The number of times to retry the job when the job fails due to an InternalServerError.

enableNetworkIsolation :: Maybe Bool

Isolates the training container. No inbound or outbound network calls can be made, except for calls between peers within a training cluster for distributed training. If network isolation is used for training jobs that are configured to use a VPC, Amazon SageMaker downloads and uploads customer data and model artifacts through the specified VPC, but the training container does not have network access.

staticHyperParameters :: Maybe (HashMap Text Text)

Specifies the values of hyperparameters that do not change for the tuning job.

enableManagedSpotTraining :: Maybe Bool

A Boolean indicating whether managed spot training is enabled (True) or not (False).

inputDataConfig :: Maybe (NonEmpty Channel)

An array of Channel objects that specify the input for the training jobs that the tuning job launches.

vpcConfig :: Maybe VpcConfig

The VpcConfig object that specifies the VPC that you want the training jobs that this hyperparameter tuning job launches to connect to. Control access to and from your training container by configuring the VPC. For more information, see Protect Training Jobs by Using an Amazon Virtual Private Cloud.

definitionName :: Maybe Text

The job definition name.

enableInterContainerTrafficEncryption :: Maybe Bool

To encrypt all communications between ML compute instances in distributed training, choose True. Encryption provides greater security for distributed training, but training might take longer. How long it takes depends on the amount of communication between compute instances, especially if you use a deep learning algorithm in distributed training.

algorithmSpecification :: HyperParameterAlgorithmSpecification

The HyperParameterAlgorithmSpecification object that specifies the resource algorithm to use for the training jobs that the tuning job launches.

roleArn :: Text

The Amazon Resource Name (ARN) of the IAM role associated with the training jobs that the tuning job launches.

outputDataConfig :: OutputDataConfig

Specifies the path to the Amazon S3 bucket where you store model artifacts from the training jobs that the tuning job launches.

resourceConfig :: ResourceConfig

The resources, including the compute instances and storage volumes, to use for the training jobs that the tuning job launches.

Storage volumes store model artifacts and incremental states. Training algorithms might also use storage volumes for scratch space. If you want Amazon SageMaker to use the storage volume to store the training data, choose File as the TrainingInputMode in the algorithm specification. For distributed training algorithms, specify an instance count greater than 1.

stoppingCondition :: StoppingCondition

Specifies a limit to how long a model hyperparameter training job can run. It also specifies how long a managed spot training job has to complete. When the job reaches the time limit, Amazon SageMaker ends the training job. Use this API to cap model training costs.

newHyperParameterTrainingJobDefinition

Arguments

:: HyperParameterAlgorithmSpecification
$sel:algorithmSpecification:HyperParameterTrainingJobDefinition'

-> Text
$sel:roleArn:HyperParameterTrainingJobDefinition'

-> OutputDataConfig
$sel:outputDataConfig:HyperParameterTrainingJobDefinition'

-> ResourceConfig
$sel:resourceConfig:HyperParameterTrainingJobDefinition'

-> StoppingCondition
$sel:stoppingCondition:HyperParameterTrainingJobDefinition'

-> HyperParameterTrainingJobDefinition

Create a value of HyperParameterTrainingJobDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tuningObjective:HyperParameterTrainingJobDefinition', hyperParameterTrainingJobDefinition_tuningObjective - Undocumented member.

$sel:checkpointConfig:HyperParameterTrainingJobDefinition', hyperParameterTrainingJobDefinition_checkpointConfig - Undocumented member.

$sel:hyperParameterRanges:HyperParameterTrainingJobDefinition', hyperParameterTrainingJobDefinition_hyperParameterRanges - Undocumented member.

$sel:retryStrategy:HyperParameterTrainingJobDefinition', hyperParameterTrainingJobDefinition_retryStrategy - The number of times to retry the job when the job fails due to an InternalServerError.

$sel:enableNetworkIsolation:HyperParameterTrainingJobDefinition', hyperParameterTrainingJobDefinition_enableNetworkIsolation - Isolates the training container. No inbound or outbound network calls can be made, except for calls between peers within a training cluster for distributed training. If network isolation is used for training jobs that are configured to use a VPC, Amazon SageMaker downloads and uploads customer data and model artifacts through the specified VPC, but the training container does not have network access.

$sel:staticHyperParameters:HyperParameterTrainingJobDefinition', hyperParameterTrainingJobDefinition_staticHyperParameters - Specifies the values of hyperparameters that do not change for the tuning job.

$sel:enableManagedSpotTraining:HyperParameterTrainingJobDefinition', hyperParameterTrainingJobDefinition_enableManagedSpotTraining - A Boolean indicating whether managed spot training is enabled (True) or not (False).

$sel:inputDataConfig:HyperParameterTrainingJobDefinition', hyperParameterTrainingJobDefinition_inputDataConfig - An array of Channel objects that specify the input for the training jobs that the tuning job launches.

$sel:vpcConfig:HyperParameterTrainingJobDefinition', hyperParameterTrainingJobDefinition_vpcConfig - The VpcConfig object that specifies the VPC that you want the training jobs that this hyperparameter tuning job launches to connect to. Control access to and from your training container by configuring the VPC. For more information, see Protect Training Jobs by Using an Amazon Virtual Private Cloud.

$sel:definitionName:HyperParameterTrainingJobDefinition', hyperParameterTrainingJobDefinition_definitionName - The job definition name.

$sel:enableInterContainerTrafficEncryption:HyperParameterTrainingJobDefinition', hyperParameterTrainingJobDefinition_enableInterContainerTrafficEncryption - To encrypt all communications between ML compute instances in distributed training, choose True. Encryption provides greater security for distributed training, but training might take longer. How long it takes depends on the amount of communication between compute instances, especially if you use a deep learning algorithm in distributed training.

$sel:algorithmSpecification:HyperParameterTrainingJobDefinition', hyperParameterTrainingJobDefinition_algorithmSpecification - The HyperParameterAlgorithmSpecification object that specifies the resource algorithm to use for the training jobs that the tuning job launches.

$sel:roleArn:HyperParameterTrainingJobDefinition', hyperParameterTrainingJobDefinition_roleArn - The Amazon Resource Name (ARN) of the IAM role associated with the training jobs that the tuning job launches.

$sel:outputDataConfig:HyperParameterTrainingJobDefinition', hyperParameterTrainingJobDefinition_outputDataConfig - Specifies the path to the Amazon S3 bucket where you store model artifacts from the training jobs that the tuning job launches.

$sel:resourceConfig:HyperParameterTrainingJobDefinition', hyperParameterTrainingJobDefinition_resourceConfig - The resources, including the compute instances and storage volumes, to use for the training jobs that the tuning job launches.

Storage volumes store model artifacts and incremental states. Training algorithms might also use storage volumes for scratch space. If you want Amazon SageMaker to use the storage volume to store the training data, choose File as the TrainingInputMode in the algorithm specification. For distributed training algorithms, specify an instance count greater than 1.

$sel:stoppingCondition:HyperParameterTrainingJobDefinition', hyperParameterTrainingJobDefinition_stoppingCondition - Specifies a limit to how long a model hyperparameter training job can run. It also specifies how long a managed spot training job has to complete. When the job reaches the time limit, Amazon SageMaker ends the training job. Use this API to cap model training costs.

hyperParameterTrainingJobDefinition_tuningObjective :: Lens' HyperParameterTrainingJobDefinition (Maybe HyperParameterTuningJobObjective)

Undocumented member.

hyperParameterTrainingJobDefinition_checkpointConfig :: Lens' HyperParameterTrainingJobDefinition (Maybe CheckpointConfig)

Undocumented member.

hyperParameterTrainingJobDefinition_hyperParameterRanges :: Lens' HyperParameterTrainingJobDefinition (Maybe ParameterRanges)

Undocumented member.

hyperParameterTrainingJobDefinition_retryStrategy :: Lens' HyperParameterTrainingJobDefinition (Maybe RetryStrategy)

The number of times to retry the job when the job fails due to an InternalServerError.

hyperParameterTrainingJobDefinition_enableNetworkIsolation :: Lens' HyperParameterTrainingJobDefinition (Maybe Bool)

Isolates the training container. No inbound or outbound network calls can be made, except for calls between peers within a training cluster for distributed training. If network isolation is used for training jobs that are configured to use a VPC, Amazon SageMaker downloads and uploads customer data and model artifacts through the specified VPC, but the training container does not have network access.

hyperParameterTrainingJobDefinition_staticHyperParameters :: Lens' HyperParameterTrainingJobDefinition (Maybe (HashMap Text Text))

Specifies the values of hyperparameters that do not change for the tuning job.

hyperParameterTrainingJobDefinition_enableManagedSpotTraining :: Lens' HyperParameterTrainingJobDefinition (Maybe Bool)

A Boolean indicating whether managed spot training is enabled (True) or not (False).

hyperParameterTrainingJobDefinition_inputDataConfig :: Lens' HyperParameterTrainingJobDefinition (Maybe (NonEmpty Channel))

An array of Channel objects that specify the input for the training jobs that the tuning job launches.

hyperParameterTrainingJobDefinition_vpcConfig :: Lens' HyperParameterTrainingJobDefinition (Maybe VpcConfig)

The VpcConfig object that specifies the VPC that you want the training jobs that this hyperparameter tuning job launches to connect to. Control access to and from your training container by configuring the VPC. For more information, see Protect Training Jobs by Using an Amazon Virtual Private Cloud.

hyperParameterTrainingJobDefinition_definitionName :: Lens' HyperParameterTrainingJobDefinition (Maybe Text)

The job definition name.

hyperParameterTrainingJobDefinition_enableInterContainerTrafficEncryption :: Lens' HyperParameterTrainingJobDefinition (Maybe Bool)

To encrypt all communications between ML compute instances in distributed training, choose True. Encryption provides greater security for distributed training, but training might take longer. How long it takes depends on the amount of communication between compute instances, especially if you use a deep learning algorithm in distributed training.

hyperParameterTrainingJobDefinition_algorithmSpecification :: Lens' HyperParameterTrainingJobDefinition HyperParameterAlgorithmSpecification

The HyperParameterAlgorithmSpecification object that specifies the resource algorithm to use for the training jobs that the tuning job launches.

hyperParameterTrainingJobDefinition_roleArn :: Lens' HyperParameterTrainingJobDefinition Text

The Amazon Resource Name (ARN) of the IAM role associated with the training jobs that the tuning job launches.

hyperParameterTrainingJobDefinition_outputDataConfig :: Lens' HyperParameterTrainingJobDefinition OutputDataConfig

Specifies the path to the Amazon S3 bucket where you store model artifacts from the training jobs that the tuning job launches.

hyperParameterTrainingJobDefinition_resourceConfig :: Lens' HyperParameterTrainingJobDefinition ResourceConfig

The resources, including the compute instances and storage volumes, to use for the training jobs that the tuning job launches.

Storage volumes store model artifacts and incremental states. Training algorithms might also use storage volumes for scratch space. If you want Amazon SageMaker to use the storage volume to store the training data, choose File as the TrainingInputMode in the algorithm specification. For distributed training algorithms, specify an instance count greater than 1.

hyperParameterTrainingJobDefinition_stoppingCondition :: Lens' HyperParameterTrainingJobDefinition StoppingCondition

Specifies a limit to how long a model hyperparameter training job can run. It also specifies how long a managed spot training job has to complete. When the job reaches the time limit, Amazon SageMaker ends the training job. Use this API to cap model training costs.