Network.AWS.SageMaker.Types.HyperParameterTrainingJobSummary

Description

Documentation

data HyperParameterTrainingJobSummary

Specifies summary information about a training job.

See: newHyperParameterTrainingJobSummary smart constructor.

Constructors

HyperParameterTrainingJobSummary'

Fields

finalHyperParameterTuningJobObjectiveMetric :: Maybe FinalHyperParameterTuningJobObjectiveMetric

The FinalHyperParameterTuningJobObjectiveMetric object that specifies the value of the objective metric of the tuning job that launched this training job.

tuningJobName :: Maybe Text

The HyperParameter tuning job that launched the training job.

failureReason :: Maybe Text

The reason that the training job failed.

objectiveStatus :: Maybe ObjectiveStatus

The status of the objective metric for the training job:

Succeeded: The final objective metric for the training job was evaluated by the hyperparameter tuning job and used in the hyperparameter tuning process.

Pending: The training job is in progress and evaluation of its final objective metric is pending.

Failed: The final objective metric for the training job was not evaluated, and was not used in the hyperparameter tuning process. This typically occurs when the training job failed or did not emit an objective metric.

trainingStartTime :: Maybe POSIX

The date and time that the training job started.

trainingJobDefinitionName :: Maybe Text

The training job definition name.

trainingEndTime :: Maybe POSIX

Specifies the time when the training job ends on training instances. You are billed for the time interval between the value of TrainingStartTime and this time. For successful jobs and stopped jobs, this is the time after model artifacts are uploaded. For failed jobs, this is the time when Amazon SageMaker detects a job failure.

trainingJobName :: Text

The name of the training job.

trainingJobArn :: Text

The Amazon Resource Name (ARN) of the training job.

creationTime :: POSIX

The date and time that the training job was created.

trainingJobStatus :: TrainingJobStatus

The status of the training job.

tunedHyperParameters :: HashMap Text Text

A list of the hyperparameters for which you specified ranges to search.

newHyperParameterTrainingJobSummary

Arguments

:: Text
$sel:trainingJobName:HyperParameterTrainingJobSummary'

-> Text
$sel:trainingJobArn:HyperParameterTrainingJobSummary'

-> UTCTime
$sel:creationTime:HyperParameterTrainingJobSummary'

-> TrainingJobStatus
$sel:trainingJobStatus:HyperParameterTrainingJobSummary'

-> HyperParameterTrainingJobSummary

Create a value of HyperParameterTrainingJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:finalHyperParameterTuningJobObjectiveMetric:HyperParameterTrainingJobSummary', hyperParameterTrainingJobSummary_finalHyperParameterTuningJobObjectiveMetric - The FinalHyperParameterTuningJobObjectiveMetric object that specifies the value of the objective metric of the tuning job that launched this training job.

$sel:tuningJobName:HyperParameterTrainingJobSummary', hyperParameterTrainingJobSummary_tuningJobName - The HyperParameter tuning job that launched the training job.

$sel:failureReason:HyperParameterTrainingJobSummary', hyperParameterTrainingJobSummary_failureReason - The reason that the training job failed.

$sel:objectiveStatus:HyperParameterTrainingJobSummary', hyperParameterTrainingJobSummary_objectiveStatus - The status of the objective metric for the training job:

Succeeded: The final objective metric for the training job was evaluated by the hyperparameter tuning job and used in the hyperparameter tuning process.

Pending: The training job is in progress and evaluation of its final objective metric is pending.

Failed: The final objective metric for the training job was not evaluated, and was not used in the hyperparameter tuning process. This typically occurs when the training job failed or did not emit an objective metric.

$sel:trainingStartTime:HyperParameterTrainingJobSummary', hyperParameterTrainingJobSummary_trainingStartTime - The date and time that the training job started.

$sel:trainingJobDefinitionName:HyperParameterTrainingJobSummary', hyperParameterTrainingJobSummary_trainingJobDefinitionName - The training job definition name.

$sel:trainingEndTime:HyperParameterTrainingJobSummary', hyperParameterTrainingJobSummary_trainingEndTime - Specifies the time when the training job ends on training instances. You are billed for the time interval between the value of TrainingStartTime and this time. For successful jobs and stopped jobs, this is the time after model artifacts are uploaded. For failed jobs, this is the time when Amazon SageMaker detects a job failure.

$sel:trainingJobName:HyperParameterTrainingJobSummary', hyperParameterTrainingJobSummary_trainingJobName - The name of the training job.

$sel:trainingJobArn:HyperParameterTrainingJobSummary', hyperParameterTrainingJobSummary_trainingJobArn - The Amazon Resource Name (ARN) of the training job.

$sel:creationTime:HyperParameterTrainingJobSummary', hyperParameterTrainingJobSummary_creationTime - The date and time that the training job was created.

$sel:trainingJobStatus:HyperParameterTrainingJobSummary', hyperParameterTrainingJobSummary_trainingJobStatus - The status of the training job.

$sel:tunedHyperParameters:HyperParameterTrainingJobSummary', hyperParameterTrainingJobSummary_tunedHyperParameters - A list of the hyperparameters for which you specified ranges to search.

hyperParameterTrainingJobSummary_finalHyperParameterTuningJobObjectiveMetric :: Lens' HyperParameterTrainingJobSummary (Maybe FinalHyperParameterTuningJobObjectiveMetric)

The FinalHyperParameterTuningJobObjectiveMetric object that specifies the value of the objective metric of the tuning job that launched this training job.

hyperParameterTrainingJobSummary_tuningJobName :: Lens' HyperParameterTrainingJobSummary (Maybe Text)

The HyperParameter tuning job that launched the training job.

hyperParameterTrainingJobSummary_failureReason :: Lens' HyperParameterTrainingJobSummary (Maybe Text)

The reason that the training job failed.

hyperParameterTrainingJobSummary_objectiveStatus :: Lens' HyperParameterTrainingJobSummary (Maybe ObjectiveStatus)

The status of the objective metric for the training job:

Succeeded: The final objective metric for the training job was evaluated by the hyperparameter tuning job and used in the hyperparameter tuning process.

Pending: The training job is in progress and evaluation of its final objective metric is pending.

Failed: The final objective metric for the training job was not evaluated, and was not used in the hyperparameter tuning process. This typically occurs when the training job failed or did not emit an objective metric.

hyperParameterTrainingJobSummary_trainingStartTime :: Lens' HyperParameterTrainingJobSummary (Maybe UTCTime)

The date and time that the training job started.

hyperParameterTrainingJobSummary_trainingJobDefinitionName :: Lens' HyperParameterTrainingJobSummary (Maybe Text)

The training job definition name.

hyperParameterTrainingJobSummary_trainingEndTime :: Lens' HyperParameterTrainingJobSummary (Maybe UTCTime)

Specifies the time when the training job ends on training instances. You are billed for the time interval between the value of TrainingStartTime and this time. For successful jobs and stopped jobs, this is the time after model artifacts are uploaded. For failed jobs, this is the time when Amazon SageMaker detects a job failure.

hyperParameterTrainingJobSummary_trainingJobName :: Lens' HyperParameterTrainingJobSummary Text

The name of the training job.

hyperParameterTrainingJobSummary_trainingJobArn :: Lens' HyperParameterTrainingJobSummary Text

The Amazon Resource Name (ARN) of the training job.

hyperParameterTrainingJobSummary_creationTime :: Lens' HyperParameterTrainingJobSummary UTCTime

The date and time that the training job was created.

hyperParameterTrainingJobSummary_trainingJobStatus :: Lens' HyperParameterTrainingJobSummary TrainingJobStatus

The status of the training job.

hyperParameterTrainingJobSummary_tunedHyperParameters :: Lens' HyperParameterTrainingJobSummary (HashMap Text Text)

A list of the hyperparameters for which you specified ranges to search.