Network.AWS.SageMaker.Types.HyperParameterTuningJobConfig

Description

Documentation

data HyperParameterTuningJobConfig

Configures a hyperparameter tuning job.

See: newHyperParameterTuningJobConfig smart constructor.

Constructors

HyperParameterTuningJobConfig'

Fields

tuningJobCompletionCriteria :: Maybe TuningJobCompletionCriteria

The tuning job's completion criteria.

parameterRanges :: Maybe ParameterRanges

The ParameterRanges object that specifies the ranges of hyperparameters that this tuning job searches.

hyperParameterTuningJobObjective :: Maybe HyperParameterTuningJobObjective

The HyperParameterTuningJobObjective object that specifies the objective metric for this tuning job.

trainingJobEarlyStoppingType :: Maybe TrainingJobEarlyStoppingType

Specifies whether to use early stopping for training jobs launched by the hyperparameter tuning job. This can be one of the following values (the default value is OFF):

OFF

Training jobs launched by the hyperparameter tuning job do not use early stopping.

AUTO

Amazon SageMaker stops training jobs launched by the hyperparameter tuning job when they are unlikely to perform better than previously completed training jobs. For more information, see Stop Training Jobs Early.

strategy :: HyperParameterTuningJobStrategyType

Specifies how hyperparameter tuning chooses the combinations of hyperparameter values to use for the training job it launches. To use the Bayesian search strategy, set this to Bayesian. To randomly search, set it to Random. For information about search strategies, see How Hyperparameter Tuning Works.

resourceLimits :: ResourceLimits

The ResourceLimits object that specifies the maximum number of training jobs and parallel training jobs for this tuning job.

newHyperParameterTuningJobConfig

Arguments

:: HyperParameterTuningJobStrategyType
$sel:strategy:HyperParameterTuningJobConfig'

-> ResourceLimits
$sel:resourceLimits:HyperParameterTuningJobConfig'

-> HyperParameterTuningJobConfig

Create a value of HyperParameterTuningJobConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tuningJobCompletionCriteria:HyperParameterTuningJobConfig', hyperParameterTuningJobConfig_tuningJobCompletionCriteria - The tuning job's completion criteria.

$sel:parameterRanges:HyperParameterTuningJobConfig', hyperParameterTuningJobConfig_parameterRanges - The ParameterRanges object that specifies the ranges of hyperparameters that this tuning job searches.

$sel:hyperParameterTuningJobObjective:HyperParameterTuningJobConfig', hyperParameterTuningJobConfig_hyperParameterTuningJobObjective - The HyperParameterTuningJobObjective object that specifies the objective metric for this tuning job.

$sel:trainingJobEarlyStoppingType:HyperParameterTuningJobConfig', hyperParameterTuningJobConfig_trainingJobEarlyStoppingType - Specifies whether to use early stopping for training jobs launched by the hyperparameter tuning job. This can be one of the following values (the default value is OFF):

OFF

Training jobs launched by the hyperparameter tuning job do not use early stopping.

AUTO

Amazon SageMaker stops training jobs launched by the hyperparameter tuning job when they are unlikely to perform better than previously completed training jobs. For more information, see Stop Training Jobs Early.

$sel:strategy:HyperParameterTuningJobConfig', hyperParameterTuningJobConfig_strategy - Specifies how hyperparameter tuning chooses the combinations of hyperparameter values to use for the training job it launches. To use the Bayesian search strategy, set this to Bayesian. To randomly search, set it to Random. For information about search strategies, see How Hyperparameter Tuning Works.

$sel:resourceLimits:HyperParameterTuningJobConfig', hyperParameterTuningJobConfig_resourceLimits - The ResourceLimits object that specifies the maximum number of training jobs and parallel training jobs for this tuning job.

hyperParameterTuningJobConfig_tuningJobCompletionCriteria :: Lens' HyperParameterTuningJobConfig (Maybe TuningJobCompletionCriteria)

The tuning job's completion criteria.

hyperParameterTuningJobConfig_parameterRanges :: Lens' HyperParameterTuningJobConfig (Maybe ParameterRanges)

The ParameterRanges object that specifies the ranges of hyperparameters that this tuning job searches.

hyperParameterTuningJobConfig_hyperParameterTuningJobObjective :: Lens' HyperParameterTuningJobConfig (Maybe HyperParameterTuningJobObjective)

The HyperParameterTuningJobObjective object that specifies the objective metric for this tuning job.

hyperParameterTuningJobConfig_trainingJobEarlyStoppingType :: Lens' HyperParameterTuningJobConfig (Maybe TrainingJobEarlyStoppingType)

Specifies whether to use early stopping for training jobs launched by the hyperparameter tuning job. This can be one of the following values (the default value is OFF):

OFF

Training jobs launched by the hyperparameter tuning job do not use early stopping.

AUTO

Amazon SageMaker stops training jobs launched by the hyperparameter tuning job when they are unlikely to perform better than previously completed training jobs. For more information, see Stop Training Jobs Early.

hyperParameterTuningJobConfig_strategy :: Lens' HyperParameterTuningJobConfig HyperParameterTuningJobStrategyType

Specifies how hyperparameter tuning chooses the combinations of hyperparameter values to use for the training job it launches. To use the Bayesian search strategy, set this to Bayesian. To randomly search, set it to Random. For information about search strategies, see How Hyperparameter Tuning Works.

hyperParameterTuningJobConfig_resourceLimits :: Lens' HyperParameterTuningJobConfig ResourceLimits

The ResourceLimits object that specifies the maximum number of training jobs and parallel training jobs for this tuning job.