Network.AWS.SageMaker.Types.HyperParameterTuningJobObjective

Description

Documentation

data HyperParameterTuningJobObjective

Defines the objective metric for a hyperparameter tuning job. Hyperparameter tuning uses the value of this metric to evaluate the training jobs it launches, and returns the training job that results in either the highest or lowest value for this metric, depending on the value you specify for the Type parameter.

See: newHyperParameterTuningJobObjective smart constructor.

Constructors

HyperParameterTuningJobObjective'

Fields

type' :: HyperParameterTuningJobObjectiveType

Whether to minimize or maximize the objective metric.

metricName :: Text

The name of the metric to use for the objective metric.

newHyperParameterTuningJobObjective

Arguments

:: HyperParameterTuningJobObjectiveType
$sel:type':HyperParameterTuningJobObjective'

-> Text
$sel:metricName:HyperParameterTuningJobObjective'

-> HyperParameterTuningJobObjective

Create a value of HyperParameterTuningJobObjective with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':HyperParameterTuningJobObjective', hyperParameterTuningJobObjective_type - Whether to minimize or maximize the objective metric.

$sel:metricName:HyperParameterTuningJobObjective', hyperParameterTuningJobObjective_metricName - The name of the metric to use for the objective metric.

hyperParameterTuningJobObjective_type :: Lens' HyperParameterTuningJobObjective HyperParameterTuningJobObjectiveType

Whether to minimize or maximize the objective metric.

hyperParameterTuningJobObjective_metricName :: Lens' HyperParameterTuningJobObjective Text

The name of the metric to use for the objective metric.