Network.AWS.SageMaker.Types.LabelCounters

Description

Documentation

data LabelCounters

Provides a breakdown of the number of objects labeled.

See: newLabelCounters smart constructor.

Constructors

LabelCounters'

Fields

machineLabeled :: Maybe Natural

The total number of objects labeled by automated data labeling.

totalLabeled :: Maybe Natural

The total number of objects labeled.

failedNonRetryableError :: Maybe Natural

The total number of objects that could not be labeled due to an error.

unlabeled :: Maybe Natural

The total number of objects not yet labeled.

humanLabeled :: Maybe Natural

The total number of objects labeled by a human worker.

newLabelCounters :: LabelCounters

Create a value of LabelCounters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:machineLabeled:LabelCounters', labelCounters_machineLabeled - The total number of objects labeled by automated data labeling.

$sel:totalLabeled:LabelCounters', labelCounters_totalLabeled - The total number of objects labeled.

$sel:failedNonRetryableError:LabelCounters', labelCounters_failedNonRetryableError - The total number of objects that could not be labeled due to an error.

$sel:unlabeled:LabelCounters', labelCounters_unlabeled - The total number of objects not yet labeled.

$sel:humanLabeled:LabelCounters', labelCounters_humanLabeled - The total number of objects labeled by a human worker.

labelCounters_machineLabeled :: Lens' LabelCounters (Maybe Natural)

The total number of objects labeled by automated data labeling.

labelCounters_totalLabeled :: Lens' LabelCounters (Maybe Natural)

The total number of objects labeled.

labelCounters_failedNonRetryableError :: Lens' LabelCounters (Maybe Natural)

The total number of objects that could not be labeled due to an error.

labelCounters_unlabeled :: Lens' LabelCounters (Maybe Natural)

The total number of objects not yet labeled.

labelCounters_humanLabeled :: Lens' LabelCounters (Maybe Natural)

The total number of objects labeled by a human worker.