Network.AWS.SageMaker.Types.LabelingJobForWorkteamSummary

Description

Documentation

data LabelingJobForWorkteamSummary

Provides summary information for a work team.

See: newLabelingJobForWorkteamSummary smart constructor.

Constructors

LabelingJobForWorkteamSummary'

Fields

labelCounters :: Maybe LabelCountersForWorkteam

Provides information about the progress of a labeling job.

labelingJobName :: Maybe Text

The name of the labeling job that the work team is assigned to.

numberOfHumanWorkersPerDataObject :: Maybe Natural

The configured number of workers per data object.

jobReferenceCode :: Text

A unique identifier for a labeling job. You can use this to refer to a specific labeling job.

workRequesterAccountId :: Text

The Amazon Web Services account ID of the account used to start the labeling job.

creationTime :: POSIX

The date and time that the labeling job was created.

newLabelingJobForWorkteamSummary

Arguments

:: Text
$sel:jobReferenceCode:LabelingJobForWorkteamSummary'

-> Text
$sel:workRequesterAccountId:LabelingJobForWorkteamSummary'

-> UTCTime
$sel:creationTime:LabelingJobForWorkteamSummary'

-> LabelingJobForWorkteamSummary

Create a value of LabelingJobForWorkteamSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:labelCounters:LabelingJobForWorkteamSummary', labelingJobForWorkteamSummary_labelCounters - Provides information about the progress of a labeling job.

$sel:labelingJobName:LabelingJobForWorkteamSummary', labelingJobForWorkteamSummary_labelingJobName - The name of the labeling job that the work team is assigned to.

$sel:numberOfHumanWorkersPerDataObject:LabelingJobForWorkteamSummary', labelingJobForWorkteamSummary_numberOfHumanWorkersPerDataObject - The configured number of workers per data object.

$sel:jobReferenceCode:LabelingJobForWorkteamSummary', labelingJobForWorkteamSummary_jobReferenceCode - A unique identifier for a labeling job. You can use this to refer to a specific labeling job.

$sel:workRequesterAccountId:LabelingJobForWorkteamSummary', labelingJobForWorkteamSummary_workRequesterAccountId - The Amazon Web Services account ID of the account used to start the labeling job.

$sel:creationTime:LabelingJobForWorkteamSummary', labelingJobForWorkteamSummary_creationTime - The date and time that the labeling job was created.

labelingJobForWorkteamSummary_labelCounters :: Lens' LabelingJobForWorkteamSummary (Maybe LabelCountersForWorkteam)

Provides information about the progress of a labeling job.

labelingJobForWorkteamSummary_labelingJobName :: Lens' LabelingJobForWorkteamSummary (Maybe Text)

The name of the labeling job that the work team is assigned to.

labelingJobForWorkteamSummary_numberOfHumanWorkersPerDataObject :: Lens' LabelingJobForWorkteamSummary (Maybe Natural)

The configured number of workers per data object.

labelingJobForWorkteamSummary_jobReferenceCode :: Lens' LabelingJobForWorkteamSummary Text

A unique identifier for a labeling job. You can use this to refer to a specific labeling job.

labelingJobForWorkteamSummary_workRequesterAccountId :: Lens' LabelingJobForWorkteamSummary Text

The Amazon Web Services account ID of the account used to start the labeling job.

labelingJobForWorkteamSummary_creationTime :: Lens' LabelingJobForWorkteamSummary UTCTime

The date and time that the labeling job was created.