Network.AWS.SageMaker.Types.LabelingJobSummary

Description

Documentation

data LabelingJobSummary

Provides summary information about a labeling job.

See: newLabelingJobSummary smart constructor.

Constructors

LabelingJobSummary'

Fields

annotationConsolidationLambdaArn :: Maybe Text

The Amazon Resource Name (ARN) of the Lambda function used to consolidate the annotations from individual workers into a label for a data object. For more information, see Annotation Consolidation.

inputConfig :: Maybe LabelingJobInputConfig

Input configuration for the labeling job.

failureReason :: Maybe Text

If the LabelingJobStatus field is Failed, this field contains a description of the error.

labelingJobOutput :: Maybe LabelingJobOutput

The location of the output produced by the labeling job.

labelingJobName :: Text

The name of the labeling job.

labelingJobArn :: Text

The Amazon Resource Name (ARN) assigned to the labeling job when it was created.

creationTime :: POSIX

The date and time that the job was created (timestamp).

lastModifiedTime :: POSIX

The date and time that the job was last modified (timestamp).

labelingJobStatus :: LabelingJobStatus

The current status of the labeling job.

labelCounters :: LabelCounters

Counts showing the progress of the labeling job.

workteamArn :: Text

The Amazon Resource Name (ARN) of the work team assigned to the job.

preHumanTaskLambdaArn :: Text

The Amazon Resource Name (ARN) of a Lambda function. The function is run before each data object is sent to a worker.

newLabelingJobSummary

Arguments

:: Text
$sel:labelingJobName:LabelingJobSummary'

-> Text
$sel:labelingJobArn:LabelingJobSummary'

-> UTCTime
$sel:creationTime:LabelingJobSummary'

-> UTCTime
$sel:lastModifiedTime:LabelingJobSummary'

-> LabelingJobStatus
$sel:labelingJobStatus:LabelingJobSummary'

-> LabelCounters
$sel:labelCounters:LabelingJobSummary'

-> Text
$sel:workteamArn:LabelingJobSummary'

-> Text
$sel:preHumanTaskLambdaArn:LabelingJobSummary'

-> LabelingJobSummary

Create a value of LabelingJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:annotationConsolidationLambdaArn:LabelingJobSummary', labelingJobSummary_annotationConsolidationLambdaArn - The Amazon Resource Name (ARN) of the Lambda function used to consolidate the annotations from individual workers into a label for a data object. For more information, see Annotation Consolidation.

$sel:inputConfig:LabelingJobSummary', labelingJobSummary_inputConfig - Input configuration for the labeling job.

$sel:failureReason:LabelingJobSummary', labelingJobSummary_failureReason - If the LabelingJobStatus field is Failed, this field contains a description of the error.

$sel:labelingJobOutput:LabelingJobSummary', labelingJobSummary_labelingJobOutput - The location of the output produced by the labeling job.

$sel:labelingJobName:LabelingJobSummary', labelingJobSummary_labelingJobName - The name of the labeling job.

$sel:labelingJobArn:LabelingJobSummary', labelingJobSummary_labelingJobArn - The Amazon Resource Name (ARN) assigned to the labeling job when it was created.

$sel:creationTime:LabelingJobSummary', labelingJobSummary_creationTime - The date and time that the job was created (timestamp).

$sel:lastModifiedTime:LabelingJobSummary', labelingJobSummary_lastModifiedTime - The date and time that the job was last modified (timestamp).

$sel:labelingJobStatus:LabelingJobSummary', labelingJobSummary_labelingJobStatus - The current status of the labeling job.

$sel:labelCounters:LabelingJobSummary', labelingJobSummary_labelCounters - Counts showing the progress of the labeling job.

$sel:workteamArn:LabelingJobSummary', labelingJobSummary_workteamArn - The Amazon Resource Name (ARN) of the work team assigned to the job.

$sel:preHumanTaskLambdaArn:LabelingJobSummary', labelingJobSummary_preHumanTaskLambdaArn - The Amazon Resource Name (ARN) of a Lambda function. The function is run before each data object is sent to a worker.

labelingJobSummary_annotationConsolidationLambdaArn :: Lens' LabelingJobSummary (Maybe Text)

The Amazon Resource Name (ARN) of the Lambda function used to consolidate the annotations from individual workers into a label for a data object. For more information, see Annotation Consolidation.

labelingJobSummary_inputConfig :: Lens' LabelingJobSummary (Maybe LabelingJobInputConfig)

Input configuration for the labeling job.

labelingJobSummary_failureReason :: Lens' LabelingJobSummary (Maybe Text)

If the LabelingJobStatus field is Failed, this field contains a description of the error.

labelingJobSummary_labelingJobOutput :: Lens' LabelingJobSummary (Maybe LabelingJobOutput)

The location of the output produced by the labeling job.

labelingJobSummary_labelingJobName :: Lens' LabelingJobSummary Text

The name of the labeling job.

labelingJobSummary_labelingJobArn :: Lens' LabelingJobSummary Text

The Amazon Resource Name (ARN) assigned to the labeling job when it was created.

labelingJobSummary_creationTime :: Lens' LabelingJobSummary UTCTime

The date and time that the job was created (timestamp).

labelingJobSummary_lastModifiedTime :: Lens' LabelingJobSummary UTCTime

The date and time that the job was last modified (timestamp).

labelingJobSummary_labelingJobStatus :: Lens' LabelingJobSummary LabelingJobStatus

The current status of the labeling job.

labelingJobSummary_labelCounters :: Lens' LabelingJobSummary LabelCounters

Counts showing the progress of the labeling job.

labelingJobSummary_workteamArn :: Lens' LabelingJobSummary Text

The Amazon Resource Name (ARN) of the work team assigned to the job.

labelingJobSummary_preHumanTaskLambdaArn :: Lens' LabelingJobSummary Text

The Amazon Resource Name (ARN) of a Lambda function. The function is run before each data object is sent to a worker.