Network.AWS.SageMaker.Types.MemberDefinition

Description

Documentation

data MemberDefinition

Defines an Amazon Cognito or your own OIDC IdP user group that is part of a work team.

See: newMemberDefinition smart constructor.

Constructors

MemberDefinition'

Fields

oidcMemberDefinition :: Maybe OidcMemberDefinition

A list user groups that exist in your OIDC Identity Provider (IdP). One to ten groups can be used to create a single private work team. When you add a user group to the list of Groups, you can add that user group to one or more private work teams. If you add a user group to a private work team, all workers in that user group are added to the work team.

cognitoMemberDefinition :: Maybe CognitoMemberDefinition

The Amazon Cognito user group that is part of the work team.

newMemberDefinition :: MemberDefinition

Create a value of MemberDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:oidcMemberDefinition:MemberDefinition', memberDefinition_oidcMemberDefinition - A list user groups that exist in your OIDC Identity Provider (IdP). One to ten groups can be used to create a single private work team. When you add a user group to the list of Groups, you can add that user group to one or more private work teams. If you add a user group to a private work team, all workers in that user group are added to the work team.

$sel:cognitoMemberDefinition:MemberDefinition', memberDefinition_cognitoMemberDefinition - The Amazon Cognito user group that is part of the work team.

memberDefinition_oidcMemberDefinition :: Lens' MemberDefinition (Maybe OidcMemberDefinition)

A list user groups that exist in your OIDC Identity Provider (IdP). One to ten groups can be used to create a single private work team. When you add a user group to the list of Groups, you can add that user group to one or more private work teams. If you add a user group to a private work team, all workers in that user group are added to the work team.

memberDefinition_cognitoMemberDefinition :: Lens' MemberDefinition (Maybe CognitoMemberDefinition)

The Amazon Cognito user group that is part of the work team.