Network.AWS.SageMaker.Types.MetadataProperties

Description

Documentation

data MetadataProperties

Metadata properties of the tracking entity, trial, or trial component.

See: newMetadataProperties smart constructor.

Constructors

MetadataProperties'

Fields

generatedBy :: Maybe Text

The entity this entity was generated by.

commitId :: Maybe Text

The commit ID.

projectId :: Maybe Text

The project ID.

repository :: Maybe Text

The repository.

newMetadataProperties :: MetadataProperties

Create a value of MetadataProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:generatedBy:MetadataProperties', metadataProperties_generatedBy - The entity this entity was generated by.

$sel:commitId:MetadataProperties', metadataProperties_commitId - The commit ID.

$sel:projectId:MetadataProperties', metadataProperties_projectId - The project ID.

$sel:repository:MetadataProperties', metadataProperties_repository - The repository.

metadataProperties_generatedBy :: Lens' MetadataProperties (Maybe Text)

The entity this entity was generated by.

metadataProperties_commitId :: Lens' MetadataProperties (Maybe Text)

The commit ID.

metadataProperties_projectId :: Lens' MetadataProperties (Maybe Text)

The project ID.

metadataProperties_repository :: Lens' MetadataProperties (Maybe Text)

The repository.