Network.AWS.SageMaker.Types.MetricDatum

Description

Documentation

data MetricDatum

Information about the metric for a candidate produced by an AutoML job.

See: newMetricDatum smart constructor.

Constructors

MetricDatum'

Fields

set :: Maybe MetricSetSource

The dataset split from which the AutoML job produced the metric.

metricName :: Maybe AutoMLMetricEnum

The name of the metric.

value :: Maybe Double

The value of the metric.

newMetricDatum :: MetricDatum

Create a value of MetricDatum with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:set:MetricDatum', metricDatum_set - The dataset split from which the AutoML job produced the metric.

$sel:metricName:MetricDatum', metricDatum_metricName - The name of the metric.

$sel:value:MetricDatum', metricDatum_value - The value of the metric.

metricDatum_set :: Lens' MetricDatum (Maybe MetricSetSource)

The dataset split from which the AutoML job produced the metric.

metricDatum_metricName :: Lens' MetricDatum (Maybe AutoMLMetricEnum)

The name of the metric.

metricDatum_value :: Lens' MetricDatum (Maybe Double)

The value of the metric.