Network.AWS.SageMaker.Types.ModelArtifacts

Description

Documentation

data ModelArtifacts

Provides information about the location that is configured for storing model artifacts.

Model artifacts are the output that results from training a model, and typically consist of trained parameters, a model defintion that describes how to compute inferences, and other metadata.

See: newModelArtifacts smart constructor.

Constructors

ModelArtifacts'

Fields

s3ModelArtifacts :: Text

The path of the S3 object that contains the model artifacts. For example, s3://bucket-name/keynameprefix/model.tar.gz.

newModelArtifacts

Arguments

:: Text
$sel:s3ModelArtifacts:ModelArtifacts'

-> ModelArtifacts

Create a value of ModelArtifacts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3ModelArtifacts:ModelArtifacts', modelArtifacts_s3ModelArtifacts - The path of the S3 object that contains the model artifacts. For example, s3://bucket-name/keynameprefix/model.tar.gz.

modelArtifacts_s3ModelArtifacts :: Lens' ModelArtifacts Text

The path of the S3 object that contains the model artifacts. For example, s3://bucket-name/keynameprefix/model.tar.gz.