Network.AWS.SageMaker.Types.ModelDeployConfig

Description

Documentation

data ModelDeployConfig

Specifies how to generate the endpoint name for an automatic one-click Autopilot model deployment.

See: newModelDeployConfig smart constructor.

Constructors

ModelDeployConfig'

Fields

endpointName :: Maybe Text

Specifies the endpoint name to use for a one-click Autopilot model deployment if the endpoint name is not generated automatically.

Specify the EndpointName if and only if you set AutoGenerateEndpointName to False; otherwise a 400 error is thrown.

autoGenerateEndpointName :: Maybe Bool

Set to True to automatically generate an endpoint name for a one-click Autopilot model deployment; set to False otherwise. The default value is False.

If you set AutoGenerateEndpointName to True, do not specify the EndpointName; otherwise a 400 error is thrown.

newModelDeployConfig :: ModelDeployConfig

Create a value of ModelDeployConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpointName:ModelDeployConfig', modelDeployConfig_endpointName - Specifies the endpoint name to use for a one-click Autopilot model deployment if the endpoint name is not generated automatically.

Specify the EndpointName if and only if you set AutoGenerateEndpointName to False; otherwise a 400 error is thrown.

$sel:autoGenerateEndpointName:ModelDeployConfig', modelDeployConfig_autoGenerateEndpointName - Set to True to automatically generate an endpoint name for a one-click Autopilot model deployment; set to False otherwise. The default value is False.

If you set AutoGenerateEndpointName to True, do not specify the EndpointName; otherwise a 400 error is thrown.

modelDeployConfig_endpointName :: Lens' ModelDeployConfig (Maybe Text)

Specifies the endpoint name to use for a one-click Autopilot model deployment if the endpoint name is not generated automatically.

Specify the EndpointName if and only if you set AutoGenerateEndpointName to False; otherwise a 400 error is thrown.

modelDeployConfig_autoGenerateEndpointName :: Lens' ModelDeployConfig (Maybe Bool)

Set to True to automatically generate an endpoint name for a one-click Autopilot model deployment; set to False otherwise. The default value is False.

If you set AutoGenerateEndpointName to True, do not specify the EndpointName; otherwise a 400 error is thrown.