Network.AWS.SageMaker.Types.ModelMetrics

Description

Documentation

data ModelMetrics

Contains metrics captured from a model.

See: newModelMetrics smart constructor.

Constructors

ModelMetrics'

Fields

bias :: Maybe Bias

Metrics that measure bais in a model.

modelDataQuality :: Maybe ModelDataQuality

Metrics that measure the quality of the input data for a model.

modelQuality :: Maybe ModelQuality

Metrics that measure the quality of a model.

explainability :: Maybe Explainability

Metrics that help explain a model.

newModelMetrics :: ModelMetrics

Create a value of ModelMetrics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bias:ModelMetrics', modelMetrics_bias - Metrics that measure bais in a model.

$sel:modelDataQuality:ModelMetrics', modelMetrics_modelDataQuality - Metrics that measure the quality of the input data for a model.

$sel:modelQuality:ModelMetrics', modelMetrics_modelQuality - Metrics that measure the quality of a model.

$sel:explainability:ModelMetrics', modelMetrics_explainability - Metrics that help explain a model.

modelMetrics_bias :: Lens' ModelMetrics (Maybe Bias)

Metrics that measure bais in a model.

modelMetrics_modelDataQuality :: Lens' ModelMetrics (Maybe ModelDataQuality)

Metrics that measure the quality of the input data for a model.

modelMetrics_modelQuality :: Lens' ModelMetrics (Maybe ModelQuality)

Metrics that measure the quality of a model.

modelMetrics_explainability :: Lens' ModelMetrics (Maybe Explainability)

Metrics that help explain a model.