Network.AWS.SageMaker.Types.ModelPackage

Description

Documentation

data ModelPackage

A versioned model that can be deployed for SageMaker inference.

See: newModelPackage smart constructor.

Constructors

ModelPackage'

Fields

creationTime :: Maybe POSIX

The time that the model package was created.

metadataProperties :: Maybe MetadataProperties

modelApprovalStatus :: Maybe ModelApprovalStatus

The approval status of the model. This can be one of the following values.

APPROVED - The model is approved

REJECTED - The model is rejected.

PENDING_MANUAL_APPROVAL - The model is waiting for manual approval.

sourceAlgorithmSpecification :: Maybe SourceAlgorithmSpecification

modelPackageName :: Maybe Text

The name of the model.

modelPackageArn :: Maybe Text

The Amazon Resource Name (ARN) of the model package.

modelMetrics :: Maybe ModelMetrics

Metrics for the model.

modelPackageDescription :: Maybe Text

The description of the model package.

createdBy :: Maybe UserContext

lastModifiedTime :: Maybe POSIX

The last time the model package was modified.

validationSpecification :: Maybe ModelPackageValidationSpecification

inferenceSpecification :: Maybe InferenceSpecification

approvalDescription :: Maybe Text

A description provided when the model approval is set.

modelPackageVersion :: Maybe Natural

The version number of a versioned model.

certifyForMarketplace :: Maybe Bool

Whether the model package is to be certified to be listed on Amazon Web Services Marketplace. For information about listing model packages on Amazon Web Services Marketplace, see List Your Algorithm or Model Package on Amazon Web Services Marketplace.

modelPackageGroupName :: Maybe Text

The model group to which the model belongs.

lastModifiedBy :: Maybe UserContext

modelPackageStatusDetails :: Maybe ModelPackageStatusDetails

tags :: Maybe [Tag]

A list of the tags associated with the model package. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

modelPackageStatus :: Maybe ModelPackageStatus

The status of the model package. This can be one of the following values.

PENDING - The model package is pending being created.

IN_PROGRESS - The model package is in the process of being created.

COMPLETED - The model package was successfully created.

FAILED - The model package failed.

DELETING - The model package is in the process of being deleted.

newModelPackage :: ModelPackage

Create a value of ModelPackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ModelPackage', modelPackage_creationTime - The time that the model package was created.

$sel:metadataProperties:ModelPackage', modelPackage_metadataProperties - Undocumented member.

$sel:modelApprovalStatus:ModelPackage', modelPackage_modelApprovalStatus - The approval status of the model. This can be one of the following values.

APPROVED - The model is approved

REJECTED - The model is rejected.

PENDING_MANUAL_APPROVAL - The model is waiting for manual approval.

$sel:sourceAlgorithmSpecification:ModelPackage', modelPackage_sourceAlgorithmSpecification - Undocumented member.

$sel:modelPackageName:ModelPackage', modelPackage_modelPackageName - The name of the model.

$sel:modelPackageArn:ModelPackage', modelPackage_modelPackageArn - The Amazon Resource Name (ARN) of the model package.

$sel:modelMetrics:ModelPackage', modelPackage_modelMetrics - Metrics for the model.

$sel:modelPackageDescription:ModelPackage', modelPackage_modelPackageDescription - The description of the model package.

$sel:createdBy:ModelPackage', modelPackage_createdBy - Undocumented member.

$sel:lastModifiedTime:ModelPackage', modelPackage_lastModifiedTime - The last time the model package was modified.

$sel:validationSpecification:ModelPackage', modelPackage_validationSpecification - Undocumented member.

$sel:inferenceSpecification:ModelPackage', modelPackage_inferenceSpecification - Undocumented member.

$sel:approvalDescription:ModelPackage', modelPackage_approvalDescription - A description provided when the model approval is set.

$sel:modelPackageVersion:ModelPackage', modelPackage_modelPackageVersion - The version number of a versioned model.

$sel:certifyForMarketplace:ModelPackage', modelPackage_certifyForMarketplace - Whether the model package is to be certified to be listed on Amazon Web Services Marketplace. For information about listing model packages on Amazon Web Services Marketplace, see List Your Algorithm or Model Package on Amazon Web Services Marketplace.

$sel:modelPackageGroupName:ModelPackage', modelPackage_modelPackageGroupName - The model group to which the model belongs.

$sel:lastModifiedBy:ModelPackage', modelPackage_lastModifiedBy - Undocumented member.

$sel:modelPackageStatusDetails:ModelPackage', modelPackage_modelPackageStatusDetails - Undocumented member.

$sel:tags:ModelPackage', modelPackage_tags - A list of the tags associated with the model package. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

$sel:modelPackageStatus:ModelPackage', modelPackage_modelPackageStatus - The status of the model package. This can be one of the following values.

PENDING - The model package is pending being created.

IN_PROGRESS - The model package is in the process of being created.

COMPLETED - The model package was successfully created.

FAILED - The model package failed.

DELETING - The model package is in the process of being deleted.

modelPackage_creationTime :: Lens' ModelPackage (Maybe UTCTime)

The time that the model package was created.

modelPackage_metadataProperties :: Lens' ModelPackage (Maybe MetadataProperties)

Undocumented member.

modelPackage_modelApprovalStatus :: Lens' ModelPackage (Maybe ModelApprovalStatus)

The approval status of the model. This can be one of the following values.

APPROVED - The model is approved

REJECTED - The model is rejected.

PENDING_MANUAL_APPROVAL - The model is waiting for manual approval.

modelPackage_sourceAlgorithmSpecification :: Lens' ModelPackage (Maybe SourceAlgorithmSpecification)

Undocumented member.

modelPackage_modelPackageName :: Lens' ModelPackage (Maybe Text)

The name of the model.

modelPackage_modelPackageArn :: Lens' ModelPackage (Maybe Text)

The Amazon Resource Name (ARN) of the model package.

modelPackage_modelMetrics :: Lens' ModelPackage (Maybe ModelMetrics)

Metrics for the model.

modelPackage_modelPackageDescription :: Lens' ModelPackage (Maybe Text)

The description of the model package.

modelPackage_createdBy :: Lens' ModelPackage (Maybe UserContext)

Undocumented member.

modelPackage_lastModifiedTime :: Lens' ModelPackage (Maybe UTCTime)

The last time the model package was modified.

modelPackage_validationSpecification :: Lens' ModelPackage (Maybe ModelPackageValidationSpecification)

Undocumented member.

modelPackage_inferenceSpecification :: Lens' ModelPackage (Maybe InferenceSpecification)

Undocumented member.

modelPackage_approvalDescription :: Lens' ModelPackage (Maybe Text)

A description provided when the model approval is set.

modelPackage_modelPackageVersion :: Lens' ModelPackage (Maybe Natural)

The version number of a versioned model.

modelPackage_certifyForMarketplace :: Lens' ModelPackage (Maybe Bool)

Whether the model package is to be certified to be listed on Amazon Web Services Marketplace. For information about listing model packages on Amazon Web Services Marketplace, see List Your Algorithm or Model Package on Amazon Web Services Marketplace.

modelPackage_modelPackageGroupName :: Lens' ModelPackage (Maybe Text)

The model group to which the model belongs.

modelPackage_lastModifiedBy :: Lens' ModelPackage (Maybe UserContext)

Undocumented member.

modelPackage_modelPackageStatusDetails :: Lens' ModelPackage (Maybe ModelPackageStatusDetails)

Undocumented member.

modelPackage_tags :: Lens' ModelPackage (Maybe [Tag])

A list of the tags associated with the model package. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

modelPackage_modelPackageStatus :: Lens' ModelPackage (Maybe ModelPackageStatus)

The status of the model package. This can be one of the following values.

PENDING - The model package is pending being created.

IN_PROGRESS - The model package is in the process of being created.

COMPLETED - The model package was successfully created.

FAILED - The model package failed.

DELETING - The model package is in the process of being deleted.