Network.AWS.SageMaker.Types.ModelPackageContainerDefinition

Description

Documentation

data ModelPackageContainerDefinition

Describes the Docker container for the model package.

See: newModelPackageContainerDefinition smart constructor.

Constructors

ModelPackageContainerDefinition'

Fields

imageDigest :: Maybe Text

An MD5 hash of the training algorithm that identifies the Docker image used for training.

modelDataUrl :: Maybe Text

The Amazon S3 path where the model artifacts, which result from model training, are stored. This path must point to a single gzip compressed tar archive (.tar.gz suffix).

The model artifacts must be in an S3 bucket that is in the same region as the model package.

containerHostname :: Maybe Text

The DNS host name for the Docker container.

environment :: Maybe (HashMap Text Text)

The environment variables to set in the Docker container. Each key and value in the Environment string to string map can have length of up to 1024. We support up to 16 entries in the map.

productId :: Maybe Text

The Amazon Web Services Marketplace product ID of the model package.

image :: Text

The Amazon EC2 Container Registry (Amazon ECR) path where inference code is stored.

If you are using your own custom algorithm instead of an algorithm provided by Amazon SageMaker, the inference code must meet Amazon SageMaker requirements. Amazon SageMaker supports both registry/repository[:tag] and registry/repository[@digest] image path formats. For more information, see Using Your Own Algorithms with Amazon SageMaker.

newModelPackageContainerDefinition

Arguments

:: Text
$sel:image:ModelPackageContainerDefinition'

-> ModelPackageContainerDefinition

Create a value of ModelPackageContainerDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageDigest:ModelPackageContainerDefinition', modelPackageContainerDefinition_imageDigest - An MD5 hash of the training algorithm that identifies the Docker image used for training.

$sel:modelDataUrl:ModelPackageContainerDefinition', modelPackageContainerDefinition_modelDataUrl - The Amazon S3 path where the model artifacts, which result from model training, are stored. This path must point to a single gzip compressed tar archive (.tar.gz suffix).

The model artifacts must be in an S3 bucket that is in the same region as the model package.

$sel:containerHostname:ModelPackageContainerDefinition', modelPackageContainerDefinition_containerHostname - The DNS host name for the Docker container.

$sel:environment:ModelPackageContainerDefinition', modelPackageContainerDefinition_environment - The environment variables to set in the Docker container. Each key and value in the Environment string to string map can have length of up to 1024. We support up to 16 entries in the map.

$sel:productId:ModelPackageContainerDefinition', modelPackageContainerDefinition_productId - The Amazon Web Services Marketplace product ID of the model package.

$sel:image:ModelPackageContainerDefinition', modelPackageContainerDefinition_image - The Amazon EC2 Container Registry (Amazon ECR) path where inference code is stored.

If you are using your own custom algorithm instead of an algorithm provided by Amazon SageMaker, the inference code must meet Amazon SageMaker requirements. Amazon SageMaker supports both registry/repository[:tag] and registry/repository[@digest] image path formats. For more information, see Using Your Own Algorithms with Amazon SageMaker.

modelPackageContainerDefinition_imageDigest :: Lens' ModelPackageContainerDefinition (Maybe Text)

An MD5 hash of the training algorithm that identifies the Docker image used for training.

modelPackageContainerDefinition_modelDataUrl :: Lens' ModelPackageContainerDefinition (Maybe Text)

The Amazon S3 path where the model artifacts, which result from model training, are stored. This path must point to a single gzip compressed tar archive (.tar.gz suffix).

The model artifacts must be in an S3 bucket that is in the same region as the model package.

modelPackageContainerDefinition_containerHostname :: Lens' ModelPackageContainerDefinition (Maybe Text)

The DNS host name for the Docker container.

modelPackageContainerDefinition_environment :: Lens' ModelPackageContainerDefinition (Maybe (HashMap Text Text))

The environment variables to set in the Docker container. Each key and value in the Environment string to string map can have length of up to 1024. We support up to 16 entries in the map.

modelPackageContainerDefinition_productId :: Lens' ModelPackageContainerDefinition (Maybe Text)

The Amazon Web Services Marketplace product ID of the model package.

modelPackageContainerDefinition_image :: Lens' ModelPackageContainerDefinition Text

The Amazon EC2 Container Registry (Amazon ECR) path where inference code is stored.

If you are using your own custom algorithm instead of an algorithm provided by Amazon SageMaker, the inference code must meet Amazon SageMaker requirements. Amazon SageMaker supports both registry/repository[:tag] and registry/repository[@digest] image path formats. For more information, see Using Your Own Algorithms with Amazon SageMaker.