Network.AWS.SageMaker.Types.MonitoringJobDefinitionSummary

Description

Documentation

data MonitoringJobDefinitionSummary

Summary information about a monitoring job.

See: newMonitoringJobDefinitionSummary smart constructor.

Constructors

MonitoringJobDefinitionSummary'

Fields

monitoringJobDefinitionName :: Text

The name of the monitoring job.

monitoringJobDefinitionArn :: Text

The Amazon Resource Name (ARN) of the monitoring job.

creationTime :: POSIX

The time that the monitoring job was created.

endpointName :: Text

The name of the endpoint that the job monitors.

newMonitoringJobDefinitionSummary

Arguments

:: Text
$sel:monitoringJobDefinitionName:MonitoringJobDefinitionSummary'

-> Text
$sel:monitoringJobDefinitionArn:MonitoringJobDefinitionSummary'

-> UTCTime
$sel:creationTime:MonitoringJobDefinitionSummary'

-> Text
$sel:endpointName:MonitoringJobDefinitionSummary'

-> MonitoringJobDefinitionSummary

Create a value of MonitoringJobDefinitionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:monitoringJobDefinitionName:MonitoringJobDefinitionSummary', monitoringJobDefinitionSummary_monitoringJobDefinitionName - The name of the monitoring job.

$sel:monitoringJobDefinitionArn:MonitoringJobDefinitionSummary', monitoringJobDefinitionSummary_monitoringJobDefinitionArn - The Amazon Resource Name (ARN) of the monitoring job.

$sel:creationTime:MonitoringJobDefinitionSummary', monitoringJobDefinitionSummary_creationTime - The time that the monitoring job was created.

$sel:endpointName:MonitoringJobDefinitionSummary', monitoringJobDefinitionSummary_endpointName - The name of the endpoint that the job monitors.

monitoringJobDefinitionSummary_monitoringJobDefinitionName :: Lens' MonitoringJobDefinitionSummary Text

The name of the monitoring job.

monitoringJobDefinitionSummary_monitoringJobDefinitionArn :: Lens' MonitoringJobDefinitionSummary Text

The Amazon Resource Name (ARN) of the monitoring job.

monitoringJobDefinitionSummary_creationTime :: Lens' MonitoringJobDefinitionSummary UTCTime

The time that the monitoring job was created.

monitoringJobDefinitionSummary_endpointName :: Lens' MonitoringJobDefinitionSummary Text

The name of the endpoint that the job monitors.