Network.AWS.SageMaker.Types.MonitoringNetworkConfig

Description

Documentation

data MonitoringNetworkConfig

The networking configuration for the monitoring job.

See: newMonitoringNetworkConfig smart constructor.

Constructors

MonitoringNetworkConfig'

Fields

vpcConfig :: Maybe VpcConfig

enableNetworkIsolation :: Maybe Bool

Whether to allow inbound and outbound network calls to and from the containers used for the monitoring job.

enableInterContainerTrafficEncryption :: Maybe Bool

Whether to encrypt all communications between the instances used for the monitoring jobs. Choose True to encrypt communications. Encryption provides greater security for distributed jobs, but the processing might take longer.

newMonitoringNetworkConfig :: MonitoringNetworkConfig

Create a value of MonitoringNetworkConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcConfig:MonitoringNetworkConfig', monitoringNetworkConfig_vpcConfig - Undocumented member.

$sel:enableNetworkIsolation:MonitoringNetworkConfig', monitoringNetworkConfig_enableNetworkIsolation - Whether to allow inbound and outbound network calls to and from the containers used for the monitoring job.

$sel:enableInterContainerTrafficEncryption:MonitoringNetworkConfig', monitoringNetworkConfig_enableInterContainerTrafficEncryption - Whether to encrypt all communications between the instances used for the monitoring jobs. Choose True to encrypt communications. Encryption provides greater security for distributed jobs, but the processing might take longer.

monitoringNetworkConfig_vpcConfig :: Lens' MonitoringNetworkConfig (Maybe VpcConfig)

Undocumented member.

monitoringNetworkConfig_enableNetworkIsolation :: Lens' MonitoringNetworkConfig (Maybe Bool)

Whether to allow inbound and outbound network calls to and from the containers used for the monitoring job.

monitoringNetworkConfig_enableInterContainerTrafficEncryption :: Lens' MonitoringNetworkConfig (Maybe Bool)

Whether to encrypt all communications between the instances used for the monitoring jobs. Choose True to encrypt communications. Encryption provides greater security for distributed jobs, but the processing might take longer.