Network.AWS.SageMaker.Types.MonitoringScheduleConfig

Description

Documentation

data MonitoringScheduleConfig

Configures the monitoring schedule and defines the monitoring job.

See: newMonitoringScheduleConfig smart constructor.

Constructors

MonitoringScheduleConfig'

Fields

monitoringType :: Maybe MonitoringType

The type of the monitoring job definition to schedule.

scheduleConfig :: Maybe ScheduleConfig

Configures the monitoring schedule.

monitoringJobDefinition :: Maybe MonitoringJobDefinition

Defines the monitoring job.

monitoringJobDefinitionName :: Maybe Text

The name of the monitoring job definition to schedule.

newMonitoringScheduleConfig :: MonitoringScheduleConfig

Create a value of MonitoringScheduleConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:monitoringType:MonitoringScheduleConfig', monitoringScheduleConfig_monitoringType - The type of the monitoring job definition to schedule.

$sel:scheduleConfig:MonitoringScheduleConfig', monitoringScheduleConfig_scheduleConfig - Configures the monitoring schedule.

$sel:monitoringJobDefinition:MonitoringScheduleConfig', monitoringScheduleConfig_monitoringJobDefinition - Defines the monitoring job.

$sel:monitoringJobDefinitionName:MonitoringScheduleConfig', monitoringScheduleConfig_monitoringJobDefinitionName - The name of the monitoring job definition to schedule.

monitoringScheduleConfig_monitoringType :: Lens' MonitoringScheduleConfig (Maybe MonitoringType)

The type of the monitoring job definition to schedule.

monitoringScheduleConfig_scheduleConfig :: Lens' MonitoringScheduleConfig (Maybe ScheduleConfig)

Configures the monitoring schedule.

monitoringScheduleConfig_monitoringJobDefinition :: Lens' MonitoringScheduleConfig (Maybe MonitoringJobDefinition)

Defines the monitoring job.

monitoringScheduleConfig_monitoringJobDefinitionName :: Lens' MonitoringScheduleConfig (Maybe Text)

The name of the monitoring job definition to schedule.