Network.AWS.SageMaker.Types.MonitoringScheduleSummary

Description

Documentation

data MonitoringScheduleSummary

Summarizes the monitoring schedule.

See: newMonitoringScheduleSummary smart constructor.

Constructors

MonitoringScheduleSummary'

Fields

monitoringType :: Maybe MonitoringType

The type of the monitoring job definition that the schedule is for.

endpointName :: Maybe Text

The name of the endpoint using the monitoring schedule.

monitoringJobDefinitionName :: Maybe Text

The name of the monitoring job definition that the schedule is for.

monitoringScheduleName :: Text

The name of the monitoring schedule.

monitoringScheduleArn :: Text

The Amazon Resource Name (ARN) of the monitoring schedule.

creationTime :: POSIX

The creation time of the monitoring schedule.

lastModifiedTime :: POSIX

The last time the monitoring schedule was modified.

monitoringScheduleStatus :: ScheduleStatus

The status of the monitoring schedule.

newMonitoringScheduleSummary

Arguments

:: Text
$sel:monitoringScheduleName:MonitoringScheduleSummary'

-> Text
$sel:monitoringScheduleArn:MonitoringScheduleSummary'

-> UTCTime
$sel:creationTime:MonitoringScheduleSummary'

-> UTCTime
$sel:lastModifiedTime:MonitoringScheduleSummary'

-> ScheduleStatus
$sel:monitoringScheduleStatus:MonitoringScheduleSummary'

-> MonitoringScheduleSummary

Create a value of MonitoringScheduleSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:monitoringType:MonitoringScheduleSummary', monitoringScheduleSummary_monitoringType - The type of the monitoring job definition that the schedule is for.

$sel:endpointName:MonitoringScheduleSummary', monitoringScheduleSummary_endpointName - The name of the endpoint using the monitoring schedule.

$sel:monitoringJobDefinitionName:MonitoringScheduleSummary', monitoringScheduleSummary_monitoringJobDefinitionName - The name of the monitoring job definition that the schedule is for.

$sel:monitoringScheduleName:MonitoringScheduleSummary', monitoringScheduleSummary_monitoringScheduleName - The name of the monitoring schedule.

$sel:monitoringScheduleArn:MonitoringScheduleSummary', monitoringScheduleSummary_monitoringScheduleArn - The Amazon Resource Name (ARN) of the monitoring schedule.

$sel:creationTime:MonitoringScheduleSummary', monitoringScheduleSummary_creationTime - The creation time of the monitoring schedule.

$sel:lastModifiedTime:MonitoringScheduleSummary', monitoringScheduleSummary_lastModifiedTime - The last time the monitoring schedule was modified.

$sel:monitoringScheduleStatus:MonitoringScheduleSummary', monitoringScheduleSummary_monitoringScheduleStatus - The status of the monitoring schedule.

monitoringScheduleSummary_monitoringType :: Lens' MonitoringScheduleSummary (Maybe MonitoringType)

The type of the monitoring job definition that the schedule is for.

monitoringScheduleSummary_endpointName :: Lens' MonitoringScheduleSummary (Maybe Text)

The name of the endpoint using the monitoring schedule.

monitoringScheduleSummary_monitoringJobDefinitionName :: Lens' MonitoringScheduleSummary (Maybe Text)

The name of the monitoring job definition that the schedule is for.

monitoringScheduleSummary_monitoringScheduleName :: Lens' MonitoringScheduleSummary Text

The name of the monitoring schedule.

monitoringScheduleSummary_monitoringScheduleArn :: Lens' MonitoringScheduleSummary Text

The Amazon Resource Name (ARN) of the monitoring schedule.

monitoringScheduleSummary_creationTime :: Lens' MonitoringScheduleSummary UTCTime

The creation time of the monitoring schedule.

monitoringScheduleSummary_lastModifiedTime :: Lens' MonitoringScheduleSummary UTCTime

The last time the monitoring schedule was modified.

monitoringScheduleSummary_monitoringScheduleStatus :: Lens' MonitoringScheduleSummary ScheduleStatus

The status of the monitoring schedule.