Network.AWS.SageMaker.Types.ObjectiveStatusCounters

Description

Documentation

data ObjectiveStatusCounters

Specifies the number of training jobs that this hyperparameter tuning job launched, categorized by the status of their objective metric. The objective metric status shows whether the final objective metric for the training job has been evaluated by the tuning job and used in the hyperparameter tuning process.

See: newObjectiveStatusCounters smart constructor.

Constructors

ObjectiveStatusCounters'

Fields

succeeded :: Maybe Natural

The number of training jobs whose final objective metric was evaluated by the hyperparameter tuning job and used in the hyperparameter tuning process.

pending :: Maybe Natural

The number of training jobs that are in progress and pending evaluation of their final objective metric.

failed :: Maybe Natural

The number of training jobs whose final objective metric was not evaluated and used in the hyperparameter tuning process. This typically occurs when the training job failed or did not emit an objective metric.

newObjectiveStatusCounters :: ObjectiveStatusCounters

Create a value of ObjectiveStatusCounters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:succeeded:ObjectiveStatusCounters', objectiveStatusCounters_succeeded - The number of training jobs whose final objective metric was evaluated by the hyperparameter tuning job and used in the hyperparameter tuning process.

$sel:pending:ObjectiveStatusCounters', objectiveStatusCounters_pending - The number of training jobs that are in progress and pending evaluation of their final objective metric.

$sel:failed:ObjectiveStatusCounters', objectiveStatusCounters_failed - The number of training jobs whose final objective metric was not evaluated and used in the hyperparameter tuning process. This typically occurs when the training job failed or did not emit an objective metric.

objectiveStatusCounters_succeeded :: Lens' ObjectiveStatusCounters (Maybe Natural)

The number of training jobs whose final objective metric was evaluated by the hyperparameter tuning job and used in the hyperparameter tuning process.

objectiveStatusCounters_pending :: Lens' ObjectiveStatusCounters (Maybe Natural)

The number of training jobs that are in progress and pending evaluation of their final objective metric.

objectiveStatusCounters_failed :: Lens' ObjectiveStatusCounters (Maybe Natural)

The number of training jobs whose final objective metric was not evaluated and used in the hyperparameter tuning process. This typically occurs when the training job failed or did not emit an objective metric.