Network.AWS.SageMaker.Types.Pipeline

Description

Documentation

data Pipeline

A SageMaker Model Building Pipeline instance.

See: newPipeline smart constructor.

Constructors

Pipeline'

Fields

pipelineArn :: Maybe Text

The Amazon Resource Name (ARN) of the pipeline.

creationTime :: Maybe POSIX

The creation time of the pipeline.

pipelineDescription :: Maybe Text

The description of the pipeline.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the role that created the pipeline.

lastRunTime :: Maybe POSIX

The time when the pipeline was last run.

pipelineDisplayName :: Maybe Text

The display name of the pipeline.

tags :: Maybe [Tag]

A list of tags that apply to the pipeline.

lastModifiedTime :: Maybe POSIX

The time that the pipeline was last modified.

pipelineStatus :: Maybe PipelineStatus

The status of the pipeline.

createdBy :: Maybe UserContext

lastModifiedBy :: Maybe UserContext

pipelineName :: Maybe Text

The name of the pipeline.

newPipeline :: Pipeline

Create a value of Pipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineArn:Pipeline', pipeline_pipelineArn - The Amazon Resource Name (ARN) of the pipeline.

$sel:creationTime:Pipeline', pipeline_creationTime - The creation time of the pipeline.

$sel:pipelineDescription:Pipeline', pipeline_pipelineDescription - The description of the pipeline.

$sel:roleArn:Pipeline', pipeline_roleArn - The Amazon Resource Name (ARN) of the role that created the pipeline.

$sel:lastRunTime:Pipeline', pipeline_lastRunTime - The time when the pipeline was last run.

$sel:pipelineDisplayName:Pipeline', pipeline_pipelineDisplayName - The display name of the pipeline.

$sel:tags:Pipeline', pipeline_tags - A list of tags that apply to the pipeline.

$sel:lastModifiedTime:Pipeline', pipeline_lastModifiedTime - The time that the pipeline was last modified.

$sel:pipelineStatus:Pipeline', pipeline_pipelineStatus - The status of the pipeline.

$sel:createdBy:Pipeline', pipeline_createdBy - Undocumented member.

$sel:lastModifiedBy:Pipeline', pipeline_lastModifiedBy - Undocumented member.

$sel:pipelineName:Pipeline', pipeline_pipelineName - The name of the pipeline.

pipeline_pipelineArn :: Lens' Pipeline (Maybe Text)

The Amazon Resource Name (ARN) of the pipeline.

pipeline_creationTime :: Lens' Pipeline (Maybe UTCTime)

The creation time of the pipeline.

pipeline_pipelineDescription :: Lens' Pipeline (Maybe Text)

The description of the pipeline.

pipeline_roleArn :: Lens' Pipeline (Maybe Text)

The Amazon Resource Name (ARN) of the role that created the pipeline.

pipeline_lastRunTime :: Lens' Pipeline (Maybe UTCTime)

The time when the pipeline was last run.

pipeline_pipelineDisplayName :: Lens' Pipeline (Maybe Text)

The display name of the pipeline.

pipeline_tags :: Lens' Pipeline (Maybe [Tag])

A list of tags that apply to the pipeline.

pipeline_lastModifiedTime :: Lens' Pipeline (Maybe UTCTime)

The time that the pipeline was last modified.

pipeline_pipelineStatus :: Lens' Pipeline (Maybe PipelineStatus)

The status of the pipeline.

pipeline_createdBy :: Lens' Pipeline (Maybe UserContext)

Undocumented member.

pipeline_lastModifiedBy :: Lens' Pipeline (Maybe UserContext)

Undocumented member.

pipeline_pipelineName :: Lens' Pipeline (Maybe Text)

The name of the pipeline.