Network.AWS.SageMaker.Types.PipelineExecution

Description

Documentation

data PipelineExecution

An execution of a pipeline.

See: newPipelineExecution smart constructor.

Constructors

PipelineExecution'

Fields

pipelineArn :: Maybe Text

The Amazon Resource Name (ARN) of the pipeline that was executed.

creationTime :: Maybe POSIX

The creation time of the pipeline execution.

pipelineExecutionDescription :: Maybe Text

The description of the pipeline execution.

pipelineParameters :: Maybe [Parameter]

Contains a list of pipeline parameters. This list can be empty.

pipelineExecutionDisplayName :: Maybe Text

The display name of the pipeline execution.

pipelineExecutionStatus :: Maybe PipelineExecutionStatus

The status of the pipeline status.

failureReason :: Maybe Text

If the execution failed, a message describing why.

pipelineExperimentConfig :: Maybe PipelineExperimentConfig

lastModifiedTime :: Maybe POSIX

The time that the pipeline execution was last modified.

createdBy :: Maybe UserContext

lastModifiedBy :: Maybe UserContext

pipelineExecutionArn :: Maybe Text

The Amazon Resource Name (ARN) of the pipeline execution.

newPipelineExecution :: PipelineExecution

Create a value of PipelineExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineArn:PipelineExecution', pipelineExecution_pipelineArn - The Amazon Resource Name (ARN) of the pipeline that was executed.

$sel:creationTime:PipelineExecution', pipelineExecution_creationTime - The creation time of the pipeline execution.

$sel:pipelineExecutionDescription:PipelineExecution', pipelineExecution_pipelineExecutionDescription - The description of the pipeline execution.

$sel:pipelineParameters:PipelineExecution', pipelineExecution_pipelineParameters - Contains a list of pipeline parameters. This list can be empty.

$sel:pipelineExecutionDisplayName:PipelineExecution', pipelineExecution_pipelineExecutionDisplayName - The display name of the pipeline execution.

$sel:pipelineExecutionStatus:PipelineExecution', pipelineExecution_pipelineExecutionStatus - The status of the pipeline status.

$sel:failureReason:PipelineExecution', pipelineExecution_failureReason - If the execution failed, a message describing why.

$sel:pipelineExperimentConfig:PipelineExecution', pipelineExecution_pipelineExperimentConfig - Undocumented member.

$sel:lastModifiedTime:PipelineExecution', pipelineExecution_lastModifiedTime - The time that the pipeline execution was last modified.

$sel:createdBy:PipelineExecution', pipelineExecution_createdBy - Undocumented member.

$sel:lastModifiedBy:PipelineExecution', pipelineExecution_lastModifiedBy - Undocumented member.

$sel:pipelineExecutionArn:PipelineExecution', pipelineExecution_pipelineExecutionArn - The Amazon Resource Name (ARN) of the pipeline execution.

pipelineExecution_pipelineArn :: Lens' PipelineExecution (Maybe Text)

The Amazon Resource Name (ARN) of the pipeline that was executed.

pipelineExecution_creationTime :: Lens' PipelineExecution (Maybe UTCTime)

The creation time of the pipeline execution.

pipelineExecution_pipelineExecutionDescription :: Lens' PipelineExecution (Maybe Text)

The description of the pipeline execution.

pipelineExecution_pipelineParameters :: Lens' PipelineExecution (Maybe [Parameter])

Contains a list of pipeline parameters. This list can be empty.

pipelineExecution_pipelineExecutionDisplayName :: Lens' PipelineExecution (Maybe Text)

The display name of the pipeline execution.

pipelineExecution_pipelineExecutionStatus :: Lens' PipelineExecution (Maybe PipelineExecutionStatus)

The status of the pipeline status.

pipelineExecution_failureReason :: Lens' PipelineExecution (Maybe Text)

If the execution failed, a message describing why.

pipelineExecution_pipelineExperimentConfig :: Lens' PipelineExecution (Maybe PipelineExperimentConfig)

Undocumented member.

pipelineExecution_lastModifiedTime :: Lens' PipelineExecution (Maybe UTCTime)

The time that the pipeline execution was last modified.

pipelineExecution_createdBy :: Lens' PipelineExecution (Maybe UserContext)

Undocumented member.

pipelineExecution_lastModifiedBy :: Lens' PipelineExecution (Maybe UserContext)

Undocumented member.

pipelineExecution_pipelineExecutionArn :: Lens' PipelineExecution (Maybe Text)

The Amazon Resource Name (ARN) of the pipeline execution.