Network.AWS.SageMaker.Types.PipelineExecutionStepMetadata

Description

Documentation

data PipelineExecutionStepMetadata

Metadata for a step execution.

See: newPipelineExecutionStepMetadata smart constructor.

Constructors

PipelineExecutionStepMetadata'

Fields

model :: Maybe ModelStepMetadata

The Amazon Resource Name (ARN) of the model that was created by this step execution.

processingJob :: Maybe ProcessingJobStepMetadata

The Amazon Resource Name (ARN) of the processing job that was run by this step execution.

condition :: Maybe ConditionStepMetadata

The outcome of the condition evaluation that was run by this step execution.

tuningJob :: Maybe TuningJobStepMetaData

The Amazon Resource Name (ARN) of the tuning job that was run by this step execution.

lambda :: Maybe LambdaStepMetadata

The Amazon Resource Name (ARN) of the Lambda function that was run by this step execution and a list of output parameters.

trainingJob :: Maybe TrainingJobStepMetadata

The Amazon Resource Name (ARN) of the training job that was run by this step execution.

callback :: Maybe CallbackStepMetadata

The URL of the Amazon SQS queue used by this step execution, the pipeline generated token, and a list of output parameters.

registerModel :: Maybe RegisterModelStepMetadata

The Amazon Resource Name (ARN) of the model package the model was registered to by this step execution.

transformJob :: Maybe TransformJobStepMetadata

The Amazon Resource Name (ARN) of the transform job that was run by this step execution.

newPipelineExecutionStepMetadata :: PipelineExecutionStepMetadata

Create a value of PipelineExecutionStepMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:model:PipelineExecutionStepMetadata', pipelineExecutionStepMetadata_model - The Amazon Resource Name (ARN) of the model that was created by this step execution.

$sel:processingJob:PipelineExecutionStepMetadata', pipelineExecutionStepMetadata_processingJob - The Amazon Resource Name (ARN) of the processing job that was run by this step execution.

$sel:condition:PipelineExecutionStepMetadata', pipelineExecutionStepMetadata_condition - The outcome of the condition evaluation that was run by this step execution.

$sel:tuningJob:PipelineExecutionStepMetadata', pipelineExecutionStepMetadata_tuningJob - The Amazon Resource Name (ARN) of the tuning job that was run by this step execution.

$sel:lambda:PipelineExecutionStepMetadata', pipelineExecutionStepMetadata_lambda - The Amazon Resource Name (ARN) of the Lambda function that was run by this step execution and a list of output parameters.

$sel:trainingJob:PipelineExecutionStepMetadata', pipelineExecutionStepMetadata_trainingJob - The Amazon Resource Name (ARN) of the training job that was run by this step execution.

$sel:callback:PipelineExecutionStepMetadata', pipelineExecutionStepMetadata_callback - The URL of the Amazon SQS queue used by this step execution, the pipeline generated token, and a list of output parameters.

$sel:registerModel:PipelineExecutionStepMetadata', pipelineExecutionStepMetadata_registerModel - The Amazon Resource Name (ARN) of the model package the model was registered to by this step execution.

$sel:transformJob:PipelineExecutionStepMetadata', pipelineExecutionStepMetadata_transformJob - The Amazon Resource Name (ARN) of the transform job that was run by this step execution.

pipelineExecutionStepMetadata_model :: Lens' PipelineExecutionStepMetadata (Maybe ModelStepMetadata)

The Amazon Resource Name (ARN) of the model that was created by this step execution.

pipelineExecutionStepMetadata_processingJob :: Lens' PipelineExecutionStepMetadata (Maybe ProcessingJobStepMetadata)

The Amazon Resource Name (ARN) of the processing job that was run by this step execution.

pipelineExecutionStepMetadata_condition :: Lens' PipelineExecutionStepMetadata (Maybe ConditionStepMetadata)

The outcome of the condition evaluation that was run by this step execution.

pipelineExecutionStepMetadata_tuningJob :: Lens' PipelineExecutionStepMetadata (Maybe TuningJobStepMetaData)

The Amazon Resource Name (ARN) of the tuning job that was run by this step execution.

pipelineExecutionStepMetadata_lambda :: Lens' PipelineExecutionStepMetadata (Maybe LambdaStepMetadata)

The Amazon Resource Name (ARN) of the Lambda function that was run by this step execution and a list of output parameters.

pipelineExecutionStepMetadata_trainingJob :: Lens' PipelineExecutionStepMetadata (Maybe TrainingJobStepMetadata)

The Amazon Resource Name (ARN) of the training job that was run by this step execution.

pipelineExecutionStepMetadata_callback :: Lens' PipelineExecutionStepMetadata (Maybe CallbackStepMetadata)

The URL of the Amazon SQS queue used by this step execution, the pipeline generated token, and a list of output parameters.

pipelineExecutionStepMetadata_registerModel :: Lens' PipelineExecutionStepMetadata (Maybe RegisterModelStepMetadata)

The Amazon Resource Name (ARN) of the model package the model was registered to by this step execution.

pipelineExecutionStepMetadata_transformJob :: Lens' PipelineExecutionStepMetadata (Maybe TransformJobStepMetadata)

The Amazon Resource Name (ARN) of the transform job that was run by this step execution.