Network.AWS.SageMaker.Types.ProcessingClusterConfig

Description

Documentation

data ProcessingClusterConfig

Configuration for the cluster used to run a processing job.

See: newProcessingClusterConfig smart constructor.

Constructors

ProcessingClusterConfig'

Fields

volumeKmsKeyId :: Maybe Text

The Amazon Web Services Key Management Service (Amazon Web Services KMS) key that Amazon SageMaker uses to encrypt data on the storage volume attached to the ML compute instance(s) that run the processing job.

Certain Nitro-based instances include local storage, dependent on the instance type. Local storage volumes are encrypted using a hardware module on the instance. You can't request a VolumeKmsKeyId when using an instance type with local storage.

For a list of instance types that support local instance storage, see Instance Store Volumes.

For more information about local instance storage encryption, see SSD Instance Store Volumes.

instanceCount :: Natural

The number of ML compute instances to use in the processing job. For distributed processing jobs, specify a value greater than 1. The default value is 1.

instanceType :: ProcessingInstanceType

The ML compute instance type for the processing job.

volumeSizeInGB :: Natural

The size of the ML storage volume in gigabytes that you want to provision. You must specify sufficient ML storage for your scenario.

Certain Nitro-based instances include local storage with a fixed total size, dependent on the instance type. When using these instances for processing, Amazon SageMaker mounts the local instance storage instead of Amazon EBS gp2 storage. You can't request a VolumeSizeInGB greater than the total size of the local instance storage.

For a list of instance types that support local instance storage, including the total size per instance type, see Instance Store Volumes.

newProcessingClusterConfig

Arguments

:: Natural
$sel:instanceCount:ProcessingClusterConfig'

-> ProcessingInstanceType
$sel:instanceType:ProcessingClusterConfig'

-> Natural
$sel:volumeSizeInGB:ProcessingClusterConfig'

-> ProcessingClusterConfig

Create a value of ProcessingClusterConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:volumeKmsKeyId:ProcessingClusterConfig', processingClusterConfig_volumeKmsKeyId - The Amazon Web Services Key Management Service (Amazon Web Services KMS) key that Amazon SageMaker uses to encrypt data on the storage volume attached to the ML compute instance(s) that run the processing job.

Certain Nitro-based instances include local storage, dependent on the instance type. Local storage volumes are encrypted using a hardware module on the instance. You can't request a VolumeKmsKeyId when using an instance type with local storage.

For a list of instance types that support local instance storage, see Instance Store Volumes.

For more information about local instance storage encryption, see SSD Instance Store Volumes.

$sel:instanceCount:ProcessingClusterConfig', processingClusterConfig_instanceCount - The number of ML compute instances to use in the processing job. For distributed processing jobs, specify a value greater than 1. The default value is 1.

$sel:instanceType:ProcessingClusterConfig', processingClusterConfig_instanceType - The ML compute instance type for the processing job.

$sel:volumeSizeInGB:ProcessingClusterConfig', processingClusterConfig_volumeSizeInGB - The size of the ML storage volume in gigabytes that you want to provision. You must specify sufficient ML storage for your scenario.

Certain Nitro-based instances include local storage with a fixed total size, dependent on the instance type. When using these instances for processing, Amazon SageMaker mounts the local instance storage instead of Amazon EBS gp2 storage. You can't request a VolumeSizeInGB greater than the total size of the local instance storage.

For a list of instance types that support local instance storage, including the total size per instance type, see Instance Store Volumes.

processingClusterConfig_volumeKmsKeyId :: Lens' ProcessingClusterConfig (Maybe Text)

The Amazon Web Services Key Management Service (Amazon Web Services KMS) key that Amazon SageMaker uses to encrypt data on the storage volume attached to the ML compute instance(s) that run the processing job.

Certain Nitro-based instances include local storage, dependent on the instance type. Local storage volumes are encrypted using a hardware module on the instance. You can't request a VolumeKmsKeyId when using an instance type with local storage.

For a list of instance types that support local instance storage, see Instance Store Volumes.

For more information about local instance storage encryption, see SSD Instance Store Volumes.

processingClusterConfig_instanceCount :: Lens' ProcessingClusterConfig Natural

The number of ML compute instances to use in the processing job. For distributed processing jobs, specify a value greater than 1. The default value is 1.

processingClusterConfig_instanceType :: Lens' ProcessingClusterConfig ProcessingInstanceType

The ML compute instance type for the processing job.

processingClusterConfig_volumeSizeInGB :: Lens' ProcessingClusterConfig Natural

The size of the ML storage volume in gigabytes that you want to provision. You must specify sufficient ML storage for your scenario.

Certain Nitro-based instances include local storage with a fixed total size, dependent on the instance type. When using these instances for processing, Amazon SageMaker mounts the local instance storage instead of Amazon EBS gp2 storage. You can't request a VolumeSizeInGB greater than the total size of the local instance storage.

For a list of instance types that support local instance storage, including the total size per instance type, see Instance Store Volumes.