Network.AWS.SageMaker.Types.ProcessingJob

Description

Documentation

data ProcessingJob

An Amazon SageMaker processing job that is used to analyze data and evaluate models. For more information, see Process Data and Evaluate Models.

See: newProcessingJob smart constructor.

Constructors

ProcessingJob'

Fields

creationTime :: Maybe POSIX

The time the processing job was created.

networkConfig :: Maybe NetworkConfig

appSpecification :: Maybe AppSpecification

processingEndTime :: Maybe POSIX

The time that the processing job ended.

roleArn :: Maybe Text

The ARN of the role used to create the processing job.

processingOutputConfig :: Maybe ProcessingOutputConfig

experimentConfig :: Maybe ExperimentConfig

exitMessage :: Maybe Text

A string, up to one KB in size, that contains metadata from the processing container when the processing job exits.

processingJobStatus :: Maybe ProcessingJobStatus

The status of the processing job.

environment :: Maybe (HashMap Text Text)

Sets the environment variables in the Docker container.

autoMLJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the AutoML job associated with this processing job.

monitoringScheduleArn :: Maybe Text

The ARN of a monitoring schedule for an endpoint associated with this processing job.

failureReason :: Maybe Text

A string, up to one KB in size, that contains the reason a processing job failed, if it failed.

tags :: Maybe [Tag]

An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

processingJobArn :: Maybe Text

The ARN of the processing job.

lastModifiedTime :: Maybe POSIX

The time the processing job was last modified.

processingInputs :: Maybe [ProcessingInput]

List of input configurations for the processing job.

processingStartTime :: Maybe POSIX

The time that the processing job started.

stoppingCondition :: Maybe ProcessingStoppingCondition

processingJobName :: Maybe Text

The name of the processing job.

processingResources :: Maybe ProcessingResources

trainingJobArn :: Maybe Text

The ARN of the training job associated with this processing job.

newProcessingJob :: ProcessingJob

Create a value of ProcessingJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ProcessingJob', processingJob_creationTime - The time the processing job was created.

$sel:networkConfig:ProcessingJob', processingJob_networkConfig - Undocumented member.

$sel:appSpecification:ProcessingJob', processingJob_appSpecification - Undocumented member.

$sel:processingEndTime:ProcessingJob', processingJob_processingEndTime - The time that the processing job ended.

$sel:roleArn:ProcessingJob', processingJob_roleArn - The ARN of the role used to create the processing job.

$sel:processingOutputConfig:ProcessingJob', processingJob_processingOutputConfig - Undocumented member.

$sel:experimentConfig:ProcessingJob', processingJob_experimentConfig - Undocumented member.

$sel:exitMessage:ProcessingJob', processingJob_exitMessage - A string, up to one KB in size, that contains metadata from the processing container when the processing job exits.

$sel:processingJobStatus:ProcessingJob', processingJob_processingJobStatus - The status of the processing job.

$sel:environment:ProcessingJob', processingJob_environment - Sets the environment variables in the Docker container.

$sel:autoMLJobArn:ProcessingJob', processingJob_autoMLJobArn - The Amazon Resource Name (ARN) of the AutoML job associated with this processing job.

$sel:monitoringScheduleArn:ProcessingJob', processingJob_monitoringScheduleArn - The ARN of a monitoring schedule for an endpoint associated with this processing job.

$sel:failureReason:ProcessingJob', processingJob_failureReason - A string, up to one KB in size, that contains the reason a processing job failed, if it failed.

$sel:tags:ProcessingJob', processingJob_tags - An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

$sel:processingJobArn:ProcessingJob', processingJob_processingJobArn - The ARN of the processing job.

$sel:lastModifiedTime:ProcessingJob', processingJob_lastModifiedTime - The time the processing job was last modified.

$sel:processingInputs:ProcessingJob', processingJob_processingInputs - List of input configurations for the processing job.

$sel:processingStartTime:ProcessingJob', processingJob_processingStartTime - The time that the processing job started.

$sel:stoppingCondition:ProcessingJob', processingJob_stoppingCondition - Undocumented member.

$sel:processingJobName:ProcessingJob', processingJob_processingJobName - The name of the processing job.

$sel:processingResources:ProcessingJob', processingJob_processingResources - Undocumented member.

$sel:trainingJobArn:ProcessingJob', processingJob_trainingJobArn - The ARN of the training job associated with this processing job.

processingJob_creationTime :: Lens' ProcessingJob (Maybe UTCTime)

The time the processing job was created.

processingJob_networkConfig :: Lens' ProcessingJob (Maybe NetworkConfig)

Undocumented member.

processingJob_appSpecification :: Lens' ProcessingJob (Maybe AppSpecification)

Undocumented member.

processingJob_processingEndTime :: Lens' ProcessingJob (Maybe UTCTime)

The time that the processing job ended.

processingJob_roleArn :: Lens' ProcessingJob (Maybe Text)

The ARN of the role used to create the processing job.

processingJob_processingOutputConfig :: Lens' ProcessingJob (Maybe ProcessingOutputConfig)

Undocumented member.

processingJob_experimentConfig :: Lens' ProcessingJob (Maybe ExperimentConfig)

Undocumented member.

processingJob_exitMessage :: Lens' ProcessingJob (Maybe Text)

A string, up to one KB in size, that contains metadata from the processing container when the processing job exits.

processingJob_processingJobStatus :: Lens' ProcessingJob (Maybe ProcessingJobStatus)

The status of the processing job.

processingJob_environment :: Lens' ProcessingJob (Maybe (HashMap Text Text))

Sets the environment variables in the Docker container.

processingJob_autoMLJobArn :: Lens' ProcessingJob (Maybe Text)

The Amazon Resource Name (ARN) of the AutoML job associated with this processing job.

processingJob_monitoringScheduleArn :: Lens' ProcessingJob (Maybe Text)

The ARN of a monitoring schedule for an endpoint associated with this processing job.

processingJob_failureReason :: Lens' ProcessingJob (Maybe Text)

A string, up to one KB in size, that contains the reason a processing job failed, if it failed.

processingJob_tags :: Lens' ProcessingJob (Maybe [Tag])

An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

processingJob_processingJobArn :: Lens' ProcessingJob (Maybe Text)

The ARN of the processing job.

processingJob_lastModifiedTime :: Lens' ProcessingJob (Maybe UTCTime)

The time the processing job was last modified.

processingJob_processingInputs :: Lens' ProcessingJob (Maybe [ProcessingInput])

List of input configurations for the processing job.

processingJob_processingStartTime :: Lens' ProcessingJob (Maybe UTCTime)

The time that the processing job started.

processingJob_stoppingCondition :: Lens' ProcessingJob (Maybe ProcessingStoppingCondition)

Undocumented member.

processingJob_processingJobName :: Lens' ProcessingJob (Maybe Text)

The name of the processing job.

processingJob_processingResources :: Lens' ProcessingJob (Maybe ProcessingResources)

Undocumented member.

processingJob_trainingJobArn :: Lens' ProcessingJob (Maybe Text)

The ARN of the training job associated with this processing job.