Network.AWS.SageMaker.Types.ProductionVariant

Description

Documentation

data ProductionVariant

Identifies a model that you want to host and the resources chosen to deploy for hosting it. If you are deploying multiple models, tell Amazon SageMaker how to distribute traffic among the models by specifying variant weights.

See: newProductionVariant smart constructor.

Constructors

ProductionVariant'

Fields

acceleratorType :: Maybe ProductionVariantAcceleratorType

The size of the Elastic Inference (EI) instance to use for the production variant. EI instances provide on-demand GPU computing for inference. For more information, see Using Elastic Inference in Amazon SageMaker.

coreDumpConfig :: Maybe ProductionVariantCoreDumpConfig

Specifies configuration for a core dump from the model container when the process crashes.

initialVariantWeight :: Maybe Double

Determines initial traffic distribution among all of the models that you specify in the endpoint configuration. The traffic to a production variant is determined by the ratio of the VariantWeight to the sum of all VariantWeight values across all ProductionVariants. If unspecified, it defaults to 1.0.

variantName :: Text

The name of the production variant.

modelName :: Text

The name of the model that you want to host. This is the name that you specified when creating the model.

initialInstanceCount :: Natural

Number of instances to launch initially.

instanceType :: ProductionVariantInstanceType

The ML compute instance type.

newProductionVariant

Arguments

:: Text
$sel:variantName:ProductionVariant'

-> Text
$sel:modelName:ProductionVariant'

-> Natural
$sel:initialInstanceCount:ProductionVariant'

-> ProductionVariantInstanceType
$sel:instanceType:ProductionVariant'

-> ProductionVariant

Create a value of ProductionVariant with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceleratorType:ProductionVariant', productionVariant_acceleratorType - The size of the Elastic Inference (EI) instance to use for the production variant. EI instances provide on-demand GPU computing for inference. For more information, see Using Elastic Inference in Amazon SageMaker.

$sel:coreDumpConfig:ProductionVariant', productionVariant_coreDumpConfig - Specifies configuration for a core dump from the model container when the process crashes.

$sel:initialVariantWeight:ProductionVariant', productionVariant_initialVariantWeight - Determines initial traffic distribution among all of the models that you specify in the endpoint configuration. The traffic to a production variant is determined by the ratio of the VariantWeight to the sum of all VariantWeight values across all ProductionVariants. If unspecified, it defaults to 1.0.

$sel:variantName:ProductionVariant', productionVariant_variantName - The name of the production variant.

$sel:modelName:ProductionVariant', productionVariant_modelName - The name of the model that you want to host. This is the name that you specified when creating the model.

$sel:initialInstanceCount:ProductionVariant', productionVariant_initialInstanceCount - Number of instances to launch initially.

$sel:instanceType:ProductionVariant', productionVariant_instanceType - The ML compute instance type.

productionVariant_acceleratorType :: Lens' ProductionVariant (Maybe ProductionVariantAcceleratorType)

The size of the Elastic Inference (EI) instance to use for the production variant. EI instances provide on-demand GPU computing for inference. For more information, see Using Elastic Inference in Amazon SageMaker.

productionVariant_coreDumpConfig :: Lens' ProductionVariant (Maybe ProductionVariantCoreDumpConfig)

Specifies configuration for a core dump from the model container when the process crashes.

productionVariant_initialVariantWeight :: Lens' ProductionVariant (Maybe Double)

Determines initial traffic distribution among all of the models that you specify in the endpoint configuration. The traffic to a production variant is determined by the ratio of the VariantWeight to the sum of all VariantWeight values across all ProductionVariants. If unspecified, it defaults to 1.0.

productionVariant_variantName :: Lens' ProductionVariant Text

The name of the production variant.

productionVariant_modelName :: Lens' ProductionVariant Text

The name of the model that you want to host. This is the name that you specified when creating the model.

productionVariant_initialInstanceCount :: Lens' ProductionVariant Natural

Number of instances to launch initially.

productionVariant_instanceType :: Lens' ProductionVariant ProductionVariantInstanceType

The ML compute instance type.