Network.AWS.SageMaker.Types.ProductionVariantSummary

Description

Documentation

data ProductionVariantSummary

Describes weight and capacities for a production variant associated with an endpoint. If you sent a request to the UpdateEndpointWeightsAndCapacities API and the endpoint status is Updating, you get different desired and current values.

See: newProductionVariantSummary smart constructor.

Constructors

ProductionVariantSummary'

Fields

deployedImages :: Maybe [DeployedImage]

An array of DeployedImage objects that specify the Amazon EC2 Container Registry paths of the inference images deployed on instances of this ProductionVariant.

desiredInstanceCount :: Maybe Natural

The number of instances requested in the UpdateEndpointWeightsAndCapacities request.

currentWeight :: Maybe Double

The weight associated with the variant.

currentInstanceCount :: Maybe Natural

The number of instances associated with the variant.

desiredWeight :: Maybe Double

The requested weight, as specified in the UpdateEndpointWeightsAndCapacities request.

variantName :: Text

The name of the variant.

newProductionVariantSummary

Arguments

:: Text
$sel:variantName:ProductionVariantSummary'

-> ProductionVariantSummary

Create a value of ProductionVariantSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deployedImages:ProductionVariantSummary', productionVariantSummary_deployedImages - An array of DeployedImage objects that specify the Amazon EC2 Container Registry paths of the inference images deployed on instances of this ProductionVariant.

$sel:desiredInstanceCount:ProductionVariantSummary', productionVariantSummary_desiredInstanceCount - The number of instances requested in the UpdateEndpointWeightsAndCapacities request.

$sel:currentWeight:ProductionVariantSummary', productionVariantSummary_currentWeight - The weight associated with the variant.

$sel:currentInstanceCount:ProductionVariantSummary', productionVariantSummary_currentInstanceCount - The number of instances associated with the variant.

$sel:desiredWeight:ProductionVariantSummary', productionVariantSummary_desiredWeight - The requested weight, as specified in the UpdateEndpointWeightsAndCapacities request.

$sel:variantName:ProductionVariantSummary', productionVariantSummary_variantName - The name of the variant.

productionVariantSummary_deployedImages :: Lens' ProductionVariantSummary (Maybe [DeployedImage])

An array of DeployedImage objects that specify the Amazon EC2 Container Registry paths of the inference images deployed on instances of this ProductionVariant.

productionVariantSummary_desiredInstanceCount :: Lens' ProductionVariantSummary (Maybe Natural)

The number of instances requested in the UpdateEndpointWeightsAndCapacities request.

productionVariantSummary_currentWeight :: Lens' ProductionVariantSummary (Maybe Double)

The weight associated with the variant.

productionVariantSummary_currentInstanceCount :: Lens' ProductionVariantSummary (Maybe Natural)

The number of instances associated with the variant.

productionVariantSummary_desiredWeight :: Lens' ProductionVariantSummary (Maybe Double)

The requested weight, as specified in the UpdateEndpointWeightsAndCapacities request.

productionVariantSummary_variantName :: Lens' ProductionVariantSummary Text

The name of the variant.