Network.AWS.SageMaker.Types.ProfilerConfigForUpdate

Description

Documentation

data ProfilerConfigForUpdate

Configuration information for updating the Debugger profile parameters, system and framework metrics configurations, and storage paths.

See: newProfilerConfigForUpdate smart constructor.

Constructors

ProfilerConfigForUpdate'

Fields

s3OutputPath :: Maybe Text

Path to Amazon S3 storage location for system and framework metrics.

profilingParameters :: Maybe (HashMap Text Text)

Configuration information for capturing framework metrics. Available key strings for different profiling options are DetailedProfilingConfig, PythonProfilingConfig, and DataLoaderProfilingConfig. The following codes are configuration structures for the ProfilingParameters parameter. To learn more about how to configure the ProfilingParameters parameter, see Use the SageMaker and Debugger Configuration API Operations to Create, Update, and Debug Your Training Job.

profilingIntervalInMilliseconds :: Maybe Integer

A time interval for capturing system metrics in milliseconds. Available values are 100, 200, 500, 1000 (1 second), 5000 (5 seconds), and 60000 (1 minute) milliseconds. The default value is 500 milliseconds.

disableProfiler :: Maybe Bool

To disable Debugger monitoring and profiling, set to True.

newProfilerConfigForUpdate :: ProfilerConfigForUpdate

Create a value of ProfilerConfigForUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3OutputPath:ProfilerConfigForUpdate', profilerConfigForUpdate_s3OutputPath - Path to Amazon S3 storage location for system and framework metrics.

$sel:profilingParameters:ProfilerConfigForUpdate', profilerConfigForUpdate_profilingParameters - Configuration information for capturing framework metrics. Available key strings for different profiling options are DetailedProfilingConfig, PythonProfilingConfig, and DataLoaderProfilingConfig. The following codes are configuration structures for the ProfilingParameters parameter. To learn more about how to configure the ProfilingParameters parameter, see Use the SageMaker and Debugger Configuration API Operations to Create, Update, and Debug Your Training Job.

$sel:profilingIntervalInMilliseconds:ProfilerConfigForUpdate', profilerConfigForUpdate_profilingIntervalInMilliseconds - A time interval for capturing system metrics in milliseconds. Available values are 100, 200, 500, 1000 (1 second), 5000 (5 seconds), and 60000 (1 minute) milliseconds. The default value is 500 milliseconds.

$sel:disableProfiler:ProfilerConfigForUpdate', profilerConfigForUpdate_disableProfiler - To disable Debugger monitoring and profiling, set to True.

profilerConfigForUpdate_s3OutputPath :: Lens' ProfilerConfigForUpdate (Maybe Text)

Path to Amazon S3 storage location for system and framework metrics.

profilerConfigForUpdate_profilingParameters :: Lens' ProfilerConfigForUpdate (Maybe (HashMap Text Text))

Configuration information for capturing framework metrics. Available key strings for different profiling options are DetailedProfilingConfig, PythonProfilingConfig, and DataLoaderProfilingConfig. The following codes are configuration structures for the ProfilingParameters parameter. To learn more about how to configure the ProfilingParameters parameter, see Use the SageMaker and Debugger Configuration API Operations to Create, Update, and Debug Your Training Job.

profilerConfigForUpdate_profilingIntervalInMilliseconds :: Lens' ProfilerConfigForUpdate (Maybe Integer)

A time interval for capturing system metrics in milliseconds. Available values are 100, 200, 500, 1000 (1 second), 5000 (5 seconds), and 60000 (1 minute) milliseconds. The default value is 500 milliseconds.

profilerConfigForUpdate_disableProfiler :: Lens' ProfilerConfigForUpdate (Maybe Bool)

To disable Debugger monitoring and profiling, set to True.