Network.AWS.SageMaker.Types.Project

Description

Documentation

data Project

The properties of a project as returned by the Search API.

See: newProject smart constructor.

Constructors

Project'

Fields

creationTime :: Maybe POSIX

A timestamp specifying when the project was created.

serviceCatalogProvisionedProductDetails :: Maybe ServiceCatalogProvisionedProductDetails

createdBy :: Maybe UserContext

Who created the project.

projectStatus :: Maybe ProjectStatus

The status of the project.

projectName :: Maybe Text

The name of the project.

serviceCatalogProvisioningDetails :: Maybe ServiceCatalogProvisioningDetails

projectId :: Maybe Text

The ID of the project.

projectArn :: Maybe Text

The Amazon Resource Name (ARN) of the project.

projectDescription :: Maybe Text

The description of the project.

tags :: Maybe [Tag]

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

newProject :: Project

Create a value of Project with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Project', project_creationTime - A timestamp specifying when the project was created.

$sel:serviceCatalogProvisionedProductDetails:Project', project_serviceCatalogProvisionedProductDetails - Undocumented member.

$sel:createdBy:Project', project_createdBy - Who created the project.

$sel:projectStatus:Project', project_projectStatus - The status of the project.

$sel:projectName:Project', project_projectName - The name of the project.

$sel:serviceCatalogProvisioningDetails:Project', project_serviceCatalogProvisioningDetails - Undocumented member.

$sel:projectId:Project', project_projectId - The ID of the project.

$sel:projectArn:Project', project_projectArn - The Amazon Resource Name (ARN) of the project.

$sel:projectDescription:Project', project_projectDescription - The description of the project.

$sel:tags:Project', project_tags - An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

project_creationTime :: Lens' Project (Maybe UTCTime)

A timestamp specifying when the project was created.

project_serviceCatalogProvisionedProductDetails :: Lens' Project (Maybe ServiceCatalogProvisionedProductDetails)

Undocumented member.

project_createdBy :: Lens' Project (Maybe UserContext)

Who created the project.

project_projectStatus :: Lens' Project (Maybe ProjectStatus)

The status of the project.

project_projectName :: Lens' Project (Maybe Text)

The name of the project.

project_serviceCatalogProvisioningDetails :: Lens' Project (Maybe ServiceCatalogProvisioningDetails)

Undocumented member.

project_projectId :: Lens' Project (Maybe Text)

The ID of the project.

project_projectArn :: Lens' Project (Maybe Text)

The Amazon Resource Name (ARN) of the project.

project_projectDescription :: Lens' Project (Maybe Text)

The description of the project.

project_tags :: Lens' Project (Maybe [Tag])

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.