Network.AWS.SageMaker.Types.RedshiftDatasetDefinition

Description

Documentation

data RedshiftDatasetDefinition

Configuration for Redshift Dataset Definition input.

See: newRedshiftDatasetDefinition smart constructor.

Constructors

RedshiftDatasetDefinition'

Fields

outputCompression :: Maybe RedshiftResultCompressionType

kmsKeyId :: Maybe Text

The Amazon Web Services Key Management Service (Amazon Web Services KMS) key that Amazon SageMaker uses to encrypt data from a Redshift execution.

clusterId :: Text

database :: Text

dbUser :: Text

queryString :: Text

clusterRoleArn :: Text

The IAM role attached to your Redshift cluster that Amazon SageMaker uses to generate datasets.

outputS3Uri :: Text

The location in Amazon S3 where the Redshift query results are stored.

outputFormat :: RedshiftResultFormat

newRedshiftDatasetDefinition

Arguments

:: Text
$sel:clusterId:RedshiftDatasetDefinition'

-> Text
$sel:database:RedshiftDatasetDefinition'

-> Text
$sel:dbUser:RedshiftDatasetDefinition'

-> Text
$sel:queryString:RedshiftDatasetDefinition'

-> Text
$sel:clusterRoleArn:RedshiftDatasetDefinition'

-> Text
$sel:outputS3Uri:RedshiftDatasetDefinition'

-> RedshiftResultFormat
$sel:outputFormat:RedshiftDatasetDefinition'

-> RedshiftDatasetDefinition

Create a value of RedshiftDatasetDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outputCompression:RedshiftDatasetDefinition', redshiftDatasetDefinition_outputCompression - Undocumented member.

$sel:kmsKeyId:RedshiftDatasetDefinition', redshiftDatasetDefinition_kmsKeyId - The Amazon Web Services Key Management Service (Amazon Web Services KMS) key that Amazon SageMaker uses to encrypt data from a Redshift execution.

$sel:clusterId:RedshiftDatasetDefinition', redshiftDatasetDefinition_clusterId - Undocumented member.

$sel:database:RedshiftDatasetDefinition', redshiftDatasetDefinition_database - Undocumented member.

$sel:dbUser:RedshiftDatasetDefinition', redshiftDatasetDefinition_dbUser - Undocumented member.

$sel:queryString:RedshiftDatasetDefinition', redshiftDatasetDefinition_queryString - Undocumented member.

$sel:clusterRoleArn:RedshiftDatasetDefinition', redshiftDatasetDefinition_clusterRoleArn - The IAM role attached to your Redshift cluster that Amazon SageMaker uses to generate datasets.

$sel:outputS3Uri:RedshiftDatasetDefinition', redshiftDatasetDefinition_outputS3Uri - The location in Amazon S3 where the Redshift query results are stored.

$sel:outputFormat:RedshiftDatasetDefinition', redshiftDatasetDefinition_outputFormat - Undocumented member.

redshiftDatasetDefinition_outputCompression :: Lens' RedshiftDatasetDefinition (Maybe RedshiftResultCompressionType)

Undocumented member.

redshiftDatasetDefinition_kmsKeyId :: Lens' RedshiftDatasetDefinition (Maybe Text)

The Amazon Web Services Key Management Service (Amazon Web Services KMS) key that Amazon SageMaker uses to encrypt data from a Redshift execution.

redshiftDatasetDefinition_clusterId :: Lens' RedshiftDatasetDefinition Text

Undocumented member.

redshiftDatasetDefinition_database :: Lens' RedshiftDatasetDefinition Text

Undocumented member.

redshiftDatasetDefinition_dbUser :: Lens' RedshiftDatasetDefinition Text

Undocumented member.

redshiftDatasetDefinition_queryString :: Lens' RedshiftDatasetDefinition Text

Undocumented member.

redshiftDatasetDefinition_clusterRoleArn :: Lens' RedshiftDatasetDefinition Text

The IAM role attached to your Redshift cluster that Amazon SageMaker uses to generate datasets.

redshiftDatasetDefinition_outputS3Uri :: Lens' RedshiftDatasetDefinition Text

The location in Amazon S3 where the Redshift query results are stored.

redshiftDatasetDefinition_outputFormat :: Lens' RedshiftDatasetDefinition RedshiftResultFormat

Undocumented member.