Network.AWS.SageMaker.Types.S3StorageConfig

Description

Documentation

data S3StorageConfig

The Amazon Simple Storage (Amazon S3) location and and security configuration for OfflineStore.

See: newS3StorageConfig smart constructor.

Constructors

S3StorageConfig'

Fields

kmsKeyId :: Maybe Text

The Amazon Web Services Key Management Service (KMS) key ID of the key used to encrypt any objects written into the OfflineStore S3 location.

The IAM roleARN that is passed as a parameter to CreateFeatureGroup must have below permissions to the KmsKeyId:

"kms:GenerateDataKey"

resolvedOutputS3Uri :: Maybe Text

The S3 path where offline records are written.

s3Uri :: Text

The S3 URI, or location in Amazon S3, of OfflineStore.

S3 URIs have a format similar to the following: s3://example-bucket/prefix/.

newS3StorageConfig

Arguments

:: Text
$sel:s3Uri:S3StorageConfig'

-> S3StorageConfig

Create a value of S3StorageConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyId:S3StorageConfig', s3StorageConfig_kmsKeyId - The Amazon Web Services Key Management Service (KMS) key ID of the key used to encrypt any objects written into the OfflineStore S3 location.

The IAM roleARN that is passed as a parameter to CreateFeatureGroup must have below permissions to the KmsKeyId:

"kms:GenerateDataKey"

$sel:resolvedOutputS3Uri:S3StorageConfig', s3StorageConfig_resolvedOutputS3Uri - The S3 path where offline records are written.

$sel:s3Uri:S3StorageConfig', s3StorageConfig_s3Uri - The S3 URI, or location in Amazon S3, of OfflineStore.

S3 URIs have a format similar to the following: s3://example-bucket/prefix/.

s3StorageConfig_kmsKeyId :: Lens' S3StorageConfig (Maybe Text)

The Amazon Web Services Key Management Service (KMS) key ID of the key used to encrypt any objects written into the OfflineStore S3 location.

The IAM roleARN that is passed as a parameter to CreateFeatureGroup must have below permissions to the KmsKeyId:

"kms:GenerateDataKey"

s3StorageConfig_resolvedOutputS3Uri :: Lens' S3StorageConfig (Maybe Text)

The S3 path where offline records are written.

s3StorageConfig_s3Uri :: Lens' S3StorageConfig Text

The S3 URI, or location in Amazon S3, of OfflineStore.

S3 URIs have a format similar to the following: s3://example-bucket/prefix/.