Network.AWS.SageMaker.Types.SearchExpression

Description

Documentation

data SearchExpression

A multi-expression that searches for the specified resource or resources in a search. All resource objects that satisfy the expression's condition are included in the search results. You must specify at least one subexpression, filter, or nested filter. A SearchExpression can contain up to twenty elements.

A SearchExpression contains the following components:

A list of Filter objects. Each filter defines a simple Boolean expression comprised of a resource property name, Boolean operator, and value.

A list of NestedFilter objects. Each nested filter defines a list of Boolean expressions using a list of resource properties. A nested filter is satisfied if a single object in the list satisfies all Boolean expressions.

A list of SearchExpression objects. A search expression object can be nested in a list of search expression objects.

A Boolean operator: And or Or.

See: newSearchExpression smart constructor.

Constructors

SearchExpression'

Fields

subExpressions :: Maybe (NonEmpty SearchExpression)

A list of search expression objects.

operator :: Maybe BooleanOperator

A Boolean operator used to evaluate the search expression. If you want every conditional statement in all lists to be satisfied for the entire search expression to be true, specify And. If only a single conditional statement needs to be true for the entire search expression to be true, specify Or. The default value is And.

filters :: Maybe (NonEmpty Filter)

A list of filter objects.

nestedFilters :: Maybe (NonEmpty NestedFilters)

A list of nested filter objects.

newSearchExpression :: SearchExpression

Create a value of SearchExpression with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subExpressions:SearchExpression', searchExpression_subExpressions - A list of search expression objects.

$sel:operator:SearchExpression', searchExpression_operator - A Boolean operator used to evaluate the search expression. If you want every conditional statement in all lists to be satisfied for the entire search expression to be true, specify And. If only a single conditional statement needs to be true for the entire search expression to be true, specify Or. The default value is And.

$sel:filters:SearchExpression', searchExpression_filters - A list of filter objects.

$sel:nestedFilters:SearchExpression', searchExpression_nestedFilters - A list of nested filter objects.

searchExpression_subExpressions :: Lens' SearchExpression (Maybe (NonEmpty SearchExpression))

A list of search expression objects.

searchExpression_operator :: Lens' SearchExpression (Maybe BooleanOperator)

A Boolean operator used to evaluate the search expression. If you want every conditional statement in all lists to be satisfied for the entire search expression to be true, specify And. If only a single conditional statement needs to be true for the entire search expression to be true, specify Or. The default value is And.

searchExpression_filters :: Lens' SearchExpression (Maybe (NonEmpty Filter))

A list of filter objects.

searchExpression_nestedFilters :: Lens' SearchExpression (Maybe (NonEmpty NestedFilters))

A list of nested filter objects.