Network.AWS.SageMaker.Types.SearchRecord

Description

Documentation

data SearchRecord

A single resource returned as part of the Search API response.

See: newSearchRecord smart constructor.

Constructors

SearchRecord'

Fields

experiment :: Maybe Experiment

The properties of an experiment.

featureGroup :: Maybe FeatureGroup

modelPackage :: Maybe ModelPackage

project :: Maybe Project

The properties of a project.

trainingJob :: Maybe TrainingJob

The properties of a training job.

endpoint :: Maybe Endpoint

pipelineExecution :: Maybe PipelineExecution

trialComponent :: Maybe TrialComponent

The properties of a trial component.

trial :: Maybe Trial

The properties of a trial.

modelPackageGroup :: Maybe ModelPackageGroup

pipeline :: Maybe Pipeline

newSearchRecord :: SearchRecord

Create a value of SearchRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:experiment:SearchRecord', searchRecord_experiment - The properties of an experiment.

$sel:featureGroup:SearchRecord', searchRecord_featureGroup - Undocumented member.

$sel:modelPackage:SearchRecord', searchRecord_modelPackage - Undocumented member.

$sel:project:SearchRecord', searchRecord_project - The properties of a project.

$sel:trainingJob:SearchRecord', searchRecord_trainingJob - The properties of a training job.

$sel:endpoint:SearchRecord', searchRecord_endpoint - Undocumented member.

$sel:pipelineExecution:SearchRecord', searchRecord_pipelineExecution - Undocumented member.

$sel:trialComponent:SearchRecord', searchRecord_trialComponent - The properties of a trial component.

$sel:trial:SearchRecord', searchRecord_trial - The properties of a trial.

$sel:modelPackageGroup:SearchRecord', searchRecord_modelPackageGroup - Undocumented member.

$sel:pipeline:SearchRecord', searchRecord_pipeline - Undocumented member.

searchRecord_experiment :: Lens' SearchRecord (Maybe Experiment)

The properties of an experiment.

searchRecord_featureGroup :: Lens' SearchRecord (Maybe FeatureGroup)

Undocumented member.

searchRecord_modelPackage :: Lens' SearchRecord (Maybe ModelPackage)

Undocumented member.

searchRecord_project :: Lens' SearchRecord (Maybe Project)

The properties of a project.

searchRecord_trainingJob :: Lens' SearchRecord (Maybe TrainingJob)

The properties of a training job.

searchRecord_endpoint :: Lens' SearchRecord (Maybe Endpoint)

Undocumented member.

searchRecord_pipelineExecution :: Lens' SearchRecord (Maybe PipelineExecution)

Undocumented member.

searchRecord_trialComponent :: Lens' SearchRecord (Maybe TrialComponent)

The properties of a trial component.

searchRecord_trial :: Lens' SearchRecord (Maybe Trial)

The properties of a trial.

searchRecord_modelPackageGroup :: Lens' SearchRecord (Maybe ModelPackageGroup)

Undocumented member.

searchRecord_pipeline :: Lens' SearchRecord (Maybe Pipeline)

Undocumented member.