Network.AWS.SageMaker.Types.StudioLifecycleConfigDetails

Description

Documentation

data StudioLifecycleConfigDetails

Details of the Studio Lifecycle Configuration.

See: newStudioLifecycleConfigDetails smart constructor.

Constructors

StudioLifecycleConfigDetails'

Fields

creationTime :: Maybe POSIX

The creation time of the Studio Lifecycle Configuration.

studioLifecycleConfigArn :: Maybe Text

The Amazon Resource Name (ARN) of the Lifecycle Configuration.

studioLifecycleConfigName :: Maybe Text

The name of the Studio Lifecycle Configuration.

studioLifecycleConfigAppType :: Maybe StudioLifecycleConfigAppType

The App type to which the Lifecycle Configuration is attached.

lastModifiedTime :: Maybe POSIX

This value is equivalent to CreationTime because Studio Lifecycle Configurations are immutable.

newStudioLifecycleConfigDetails :: StudioLifecycleConfigDetails

Create a value of StudioLifecycleConfigDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:StudioLifecycleConfigDetails', studioLifecycleConfigDetails_creationTime - The creation time of the Studio Lifecycle Configuration.

$sel:studioLifecycleConfigArn:StudioLifecycleConfigDetails', studioLifecycleConfigDetails_studioLifecycleConfigArn - The Amazon Resource Name (ARN) of the Lifecycle Configuration.

$sel:studioLifecycleConfigName:StudioLifecycleConfigDetails', studioLifecycleConfigDetails_studioLifecycleConfigName - The name of the Studio Lifecycle Configuration.

$sel:studioLifecycleConfigAppType:StudioLifecycleConfigDetails', studioLifecycleConfigDetails_studioLifecycleConfigAppType - The App type to which the Lifecycle Configuration is attached.

$sel:lastModifiedTime:StudioLifecycleConfigDetails', studioLifecycleConfigDetails_lastModifiedTime - This value is equivalent to CreationTime because Studio Lifecycle Configurations are immutable.

studioLifecycleConfigDetails_creationTime :: Lens' StudioLifecycleConfigDetails (Maybe UTCTime)

The creation time of the Studio Lifecycle Configuration.

studioLifecycleConfigDetails_studioLifecycleConfigArn :: Lens' StudioLifecycleConfigDetails (Maybe Text)

The Amazon Resource Name (ARN) of the Lifecycle Configuration.

studioLifecycleConfigDetails_studioLifecycleConfigName :: Lens' StudioLifecycleConfigDetails (Maybe Text)

The name of the Studio Lifecycle Configuration.

studioLifecycleConfigDetails_studioLifecycleConfigAppType :: Lens' StudioLifecycleConfigDetails (Maybe StudioLifecycleConfigAppType)

The App type to which the Lifecycle Configuration is attached.

studioLifecycleConfigDetails_lastModifiedTime :: Lens' StudioLifecycleConfigDetails (Maybe UTCTime)

This value is equivalent to CreationTime because Studio Lifecycle Configurations are immutable.