Network.AWS.SageMaker.Types.SubscribedWorkteam

Description

Documentation

data SubscribedWorkteam

Describes a work team of a vendor that does the a labelling job.

See: newSubscribedWorkteam smart constructor.

Constructors

SubscribedWorkteam'

Fields

marketplaceTitle :: Maybe Text

The title of the service provided by the vendor in the Amazon Marketplace.

sellerName :: Maybe Text

The name of the vendor in the Amazon Marketplace.

listingId :: Maybe Text

Marketplace product listing ID.

marketplaceDescription :: Maybe Text

The description of the vendor from the Amazon Marketplace.

workteamArn :: Text

The Amazon Resource Name (ARN) of the vendor that you have subscribed.

newSubscribedWorkteam

Arguments

:: Text
$sel:workteamArn:SubscribedWorkteam'

-> SubscribedWorkteam

Create a value of SubscribedWorkteam with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marketplaceTitle:SubscribedWorkteam', subscribedWorkteam_marketplaceTitle - The title of the service provided by the vendor in the Amazon Marketplace.

$sel:sellerName:SubscribedWorkteam', subscribedWorkteam_sellerName - The name of the vendor in the Amazon Marketplace.

$sel:listingId:SubscribedWorkteam', subscribedWorkteam_listingId - Marketplace product listing ID.

$sel:marketplaceDescription:SubscribedWorkteam', subscribedWorkteam_marketplaceDescription - The description of the vendor from the Amazon Marketplace.

$sel:workteamArn:SubscribedWorkteam', subscribedWorkteam_workteamArn - The Amazon Resource Name (ARN) of the vendor that you have subscribed.

subscribedWorkteam_marketplaceTitle :: Lens' SubscribedWorkteam (Maybe Text)

The title of the service provided by the vendor in the Amazon Marketplace.

subscribedWorkteam_sellerName :: Lens' SubscribedWorkteam (Maybe Text)

The name of the vendor in the Amazon Marketplace.

subscribedWorkteam_listingId :: Lens' SubscribedWorkteam (Maybe Text)

Marketplace product listing ID.

subscribedWorkteam_marketplaceDescription :: Lens' SubscribedWorkteam (Maybe Text)

The description of the vendor from the Amazon Marketplace.

subscribedWorkteam_workteamArn :: Lens' SubscribedWorkteam Text

The Amazon Resource Name (ARN) of the vendor that you have subscribed.