Network.AWS.SageMaker.Types.TrainingJob

Description

Documentation

data TrainingJob

Contains information about a training job.

See: newTrainingJob smart constructor.

Constructors

TrainingJob'

Fields

vpcConfig :: Maybe VpcConfig

A VpcConfig object that specifies the VPC that this training job has access to. For more information, see Protect Training Jobs by Using an Amazon Virtual Private Cloud.

enableManagedSpotTraining :: Maybe Bool

When true, enables managed spot training using Amazon EC2 Spot instances to run training jobs instead of on-demand instances. For more information, see Managed Spot Training.

creationTime :: Maybe POSIX

A timestamp that indicates when the training job was created.

labelingJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the labeling job.

hyperParameters :: Maybe (HashMap Text Text)

Algorithm-specific parameters.

debugRuleConfigurations :: Maybe [DebugRuleConfiguration]

Information about the debug rule configuration.

inputDataConfig :: Maybe (NonEmpty Channel)

An array of Channel objects that describes each data input channel.

trainingTimeInSeconds :: Maybe Natural

The training time in seconds.

roleArn :: Maybe Text

The Amazon Web Services Identity and Access Management (IAM) role configured for the training job.

experimentConfig :: Maybe ExperimentConfig

enableNetworkIsolation :: Maybe Bool

If the TrainingJob was created with network isolation, the value is set to true. If network isolation is enabled, nodes can't communicate beyond the VPC they run in.

enableInterContainerTrafficEncryption :: Maybe Bool

To encrypt all communications between ML compute instances in distributed training, choose True. Encryption provides greater security for distributed training, but training might take longer. How long it takes depends on the amount of communication between compute instances, especially if you use a deep learning algorithm in distributed training.

trainingJobName :: Maybe Text

The name of the training job.

checkpointConfig :: Maybe CheckpointConfig

tuningJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the associated hyperparameter tuning job if the training job was launched by a hyperparameter tuning job.

outputDataConfig :: Maybe OutputDataConfig

The S3 path where model artifacts that you configured when creating the job are stored. Amazon SageMaker creates subfolders for model artifacts.

environment :: Maybe (HashMap Text Text)

The environment variables to set in the Docker container.

modelArtifacts :: Maybe ModelArtifacts

Information about the Amazon S3 location that is configured for storing model artifacts.

finalMetricDataList :: Maybe [MetricData]

A list of final metric values that are set when the training job completes. Used only if the training job was configured to use metrics.

secondaryStatusTransitions :: Maybe [SecondaryStatusTransition]

A history of all of the secondary statuses that the training job has transitioned through.

autoMLJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the job.

failureReason :: Maybe Text

If the training job failed, the reason it failed.

tags :: Maybe [Tag]

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

lastModifiedTime :: Maybe POSIX

A timestamp that indicates when the status of the training job was last modified.

secondaryStatus :: Maybe SecondaryStatus

Provides detailed information about the state of the training job. For detailed information about the secondary status of the training job, see StatusMessage under SecondaryStatusTransition.

Amazon SageMaker provides primary statuses and secondary statuses that apply to each of them:

InProgress

- Starting - Starting the training job.

Downloading - An optional stage for algorithms that support File training input mode. It indicates that data is being downloaded to the ML storage volumes.

Training - Training is in progress.

Uploading - Training is complete and the model artifacts are being uploaded to the S3 location.

Completed

- Completed - The training job has completed.

Failed

- Failed - The training job has failed. The reason for the failure is returned in the FailureReason field of DescribeTrainingJobResponse.

Stopped

- MaxRuntimeExceeded - The job stopped because it exceeded the maximum allowed runtime.

Stopped - The training job has stopped.

Stopping

- Stopping - Stopping the training job.

Valid values for SecondaryStatus are subject to change.

We no longer support the following secondary statuses:

LaunchingMLInstances

PreparingTrainingStack

DownloadingTrainingImage

tensorBoardOutputConfig :: Maybe TensorBoardOutputConfig

debugRuleEvaluationStatuses :: Maybe [DebugRuleEvaluationStatus]

Information about the evaluation status of the rules for the training job.

stoppingCondition :: Maybe StoppingCondition

Specifies a limit to how long a model training job can run. It also specifies how long a managed Spot training job has to complete. When the job reaches the time limit, Amazon SageMaker ends the training job. Use this API to cap model training costs.

To stop a job, Amazon SageMaker sends the algorithm the SIGTERM signal, which delays job termination for 120 seconds. Algorithms can use this 120-second window to save the model artifacts, so the results of training are not lost.

trainingJobStatus :: Maybe TrainingJobStatus

The status of the training job.

Training job statuses are:

InProgress - The training is in progress.

Completed - The training job has completed.

Failed - The training job has failed. To see the reason for the failure, see the FailureReason field in the response to a DescribeTrainingJobResponse call.

Stopping - The training job is stopping.

Stopped - The training job has stopped.

For more detailed information, see SecondaryStatus.

retryStrategy :: Maybe RetryStrategy

The number of times to retry the job when the job fails due to an InternalServerError.

resourceConfig :: Maybe ResourceConfig

Resources, including ML compute instances and ML storage volumes, that are configured for model training.

debugHookConfig :: Maybe DebugHookConfig

billableTimeInSeconds :: Maybe Natural

The billable time in seconds.

trainingStartTime :: Maybe POSIX

Indicates the time when the training job starts on training instances. You are billed for the time interval between this time and the value of TrainingEndTime. The start time in CloudWatch Logs might be later than this time. The difference is due to the time it takes to download the training data and to the size of the training container.

trainingJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the training job.

trainingEndTime :: Maybe POSIX

Indicates the time when the training job ends on training instances. You are billed for the time interval between the value of TrainingStartTime and this time. For successful jobs and stopped jobs, this is the time after model artifacts are uploaded. For failed jobs, this is the time when Amazon SageMaker detects a job failure.

algorithmSpecification :: Maybe AlgorithmSpecification

Information about the algorithm used for training, and algorithm metadata.

newTrainingJob :: TrainingJob

Create a value of TrainingJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcConfig:TrainingJob', trainingJob_vpcConfig - A VpcConfig object that specifies the VPC that this training job has access to. For more information, see Protect Training Jobs by Using an Amazon Virtual Private Cloud.

$sel:enableManagedSpotTraining:TrainingJob', trainingJob_enableManagedSpotTraining - When true, enables managed spot training using Amazon EC2 Spot instances to run training jobs instead of on-demand instances. For more information, see Managed Spot Training.

$sel:creationTime:TrainingJob', trainingJob_creationTime - A timestamp that indicates when the training job was created.

$sel:labelingJobArn:TrainingJob', trainingJob_labelingJobArn - The Amazon Resource Name (ARN) of the labeling job.

$sel:hyperParameters:TrainingJob', trainingJob_hyperParameters - Algorithm-specific parameters.

$sel:debugRuleConfigurations:TrainingJob', trainingJob_debugRuleConfigurations - Information about the debug rule configuration.

$sel:inputDataConfig:TrainingJob', trainingJob_inputDataConfig - An array of Channel objects that describes each data input channel.

$sel:trainingTimeInSeconds:TrainingJob', trainingJob_trainingTimeInSeconds - The training time in seconds.

$sel:roleArn:TrainingJob', trainingJob_roleArn - The Amazon Web Services Identity and Access Management (IAM) role configured for the training job.

$sel:experimentConfig:TrainingJob', trainingJob_experimentConfig - Undocumented member.

$sel:enableNetworkIsolation:TrainingJob', trainingJob_enableNetworkIsolation - If the TrainingJob was created with network isolation, the value is set to true. If network isolation is enabled, nodes can't communicate beyond the VPC they run in.

$sel:enableInterContainerTrafficEncryption:TrainingJob', trainingJob_enableInterContainerTrafficEncryption - To encrypt all communications between ML compute instances in distributed training, choose True. Encryption provides greater security for distributed training, but training might take longer. How long it takes depends on the amount of communication between compute instances, especially if you use a deep learning algorithm in distributed training.

$sel:trainingJobName:TrainingJob', trainingJob_trainingJobName - The name of the training job.

$sel:checkpointConfig:TrainingJob', trainingJob_checkpointConfig - Undocumented member.

$sel:tuningJobArn:TrainingJob', trainingJob_tuningJobArn - The Amazon Resource Name (ARN) of the associated hyperparameter tuning job if the training job was launched by a hyperparameter tuning job.

$sel:outputDataConfig:TrainingJob', trainingJob_outputDataConfig - The S3 path where model artifacts that you configured when creating the job are stored. Amazon SageMaker creates subfolders for model artifacts.

$sel:environment:TrainingJob', trainingJob_environment - The environment variables to set in the Docker container.

$sel:modelArtifacts:TrainingJob', trainingJob_modelArtifacts - Information about the Amazon S3 location that is configured for storing model artifacts.

$sel:finalMetricDataList:TrainingJob', trainingJob_finalMetricDataList - A list of final metric values that are set when the training job completes. Used only if the training job was configured to use metrics.

$sel:secondaryStatusTransitions:TrainingJob', trainingJob_secondaryStatusTransitions - A history of all of the secondary statuses that the training job has transitioned through.

$sel:autoMLJobArn:TrainingJob', trainingJob_autoMLJobArn - The Amazon Resource Name (ARN) of the job.

$sel:failureReason:TrainingJob', trainingJob_failureReason - If the training job failed, the reason it failed.

$sel:tags:TrainingJob', trainingJob_tags - An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

$sel:lastModifiedTime:TrainingJob', trainingJob_lastModifiedTime - A timestamp that indicates when the status of the training job was last modified.

$sel:secondaryStatus:TrainingJob', trainingJob_secondaryStatus - Provides detailed information about the state of the training job. For detailed information about the secondary status of the training job, see StatusMessage under SecondaryStatusTransition.

Amazon SageMaker provides primary statuses and secondary statuses that apply to each of them:

InProgress

- Starting - Starting the training job.

Downloading - An optional stage for algorithms that support File training input mode. It indicates that data is being downloaded to the ML storage volumes.

Training - Training is in progress.

Uploading - Training is complete and the model artifacts are being uploaded to the S3 location.

Completed

- Completed - The training job has completed.

Failed

- Failed - The training job has failed. The reason for the failure is returned in the FailureReason field of DescribeTrainingJobResponse.

Stopped

- MaxRuntimeExceeded - The job stopped because it exceeded the maximum allowed runtime.

Stopped - The training job has stopped.

Stopping

- Stopping - Stopping the training job.

Valid values for SecondaryStatus are subject to change.

We no longer support the following secondary statuses:

LaunchingMLInstances

PreparingTrainingStack

DownloadingTrainingImage

$sel:tensorBoardOutputConfig:TrainingJob', trainingJob_tensorBoardOutputConfig - Undocumented member.

$sel:debugRuleEvaluationStatuses:TrainingJob', trainingJob_debugRuleEvaluationStatuses - Information about the evaluation status of the rules for the training job.

$sel:stoppingCondition:TrainingJob', trainingJob_stoppingCondition - Specifies a limit to how long a model training job can run. It also specifies how long a managed Spot training job has to complete. When the job reaches the time limit, Amazon SageMaker ends the training job. Use this API to cap model training costs.

To stop a job, Amazon SageMaker sends the algorithm the SIGTERM signal, which delays job termination for 120 seconds. Algorithms can use this 120-second window to save the model artifacts, so the results of training are not lost.

$sel:trainingJobStatus:TrainingJob', trainingJob_trainingJobStatus - The status of the training job.

Training job statuses are:

InProgress - The training is in progress.

Completed - The training job has completed.

Failed - The training job has failed. To see the reason for the failure, see the FailureReason field in the response to a DescribeTrainingJobResponse call.

Stopping - The training job is stopping.

Stopped - The training job has stopped.

For more detailed information, see SecondaryStatus.

$sel:retryStrategy:TrainingJob', trainingJob_retryStrategy - The number of times to retry the job when the job fails due to an InternalServerError.

$sel:resourceConfig:TrainingJob', trainingJob_resourceConfig - Resources, including ML compute instances and ML storage volumes, that are configured for model training.

$sel:debugHookConfig:TrainingJob', trainingJob_debugHookConfig - Undocumented member.

$sel:billableTimeInSeconds:TrainingJob', trainingJob_billableTimeInSeconds - The billable time in seconds.

$sel:trainingStartTime:TrainingJob', trainingJob_trainingStartTime - Indicates the time when the training job starts on training instances. You are billed for the time interval between this time and the value of TrainingEndTime. The start time in CloudWatch Logs might be later than this time. The difference is due to the time it takes to download the training data and to the size of the training container.

$sel:trainingJobArn:TrainingJob', trainingJob_trainingJobArn - The Amazon Resource Name (ARN) of the training job.

$sel:trainingEndTime:TrainingJob', trainingJob_trainingEndTime - Indicates the time when the training job ends on training instances. You are billed for the time interval between the value of TrainingStartTime and this time. For successful jobs and stopped jobs, this is the time after model artifacts are uploaded. For failed jobs, this is the time when Amazon SageMaker detects a job failure.

$sel:algorithmSpecification:TrainingJob', trainingJob_algorithmSpecification - Information about the algorithm used for training, and algorithm metadata.

trainingJob_vpcConfig :: Lens' TrainingJob (Maybe VpcConfig)

A VpcConfig object that specifies the VPC that this training job has access to. For more information, see Protect Training Jobs by Using an Amazon Virtual Private Cloud.

trainingJob_enableManagedSpotTraining :: Lens' TrainingJob (Maybe Bool)

When true, enables managed spot training using Amazon EC2 Spot instances to run training jobs instead of on-demand instances. For more information, see Managed Spot Training.

trainingJob_creationTime :: Lens' TrainingJob (Maybe UTCTime)

A timestamp that indicates when the training job was created.

trainingJob_labelingJobArn :: Lens' TrainingJob (Maybe Text)

The Amazon Resource Name (ARN) of the labeling job.

trainingJob_hyperParameters :: Lens' TrainingJob (Maybe (HashMap Text Text))

Algorithm-specific parameters.

trainingJob_debugRuleConfigurations :: Lens' TrainingJob (Maybe [DebugRuleConfiguration])

Information about the debug rule configuration.

trainingJob_inputDataConfig :: Lens' TrainingJob (Maybe (NonEmpty Channel))

An array of Channel objects that describes each data input channel.

trainingJob_trainingTimeInSeconds :: Lens' TrainingJob (Maybe Natural)

The training time in seconds.

trainingJob_roleArn :: Lens' TrainingJob (Maybe Text)

The Amazon Web Services Identity and Access Management (IAM) role configured for the training job.

trainingJob_experimentConfig :: Lens' TrainingJob (Maybe ExperimentConfig)

Undocumented member.

trainingJob_enableNetworkIsolation :: Lens' TrainingJob (Maybe Bool)

If the TrainingJob was created with network isolation, the value is set to true. If network isolation is enabled, nodes can't communicate beyond the VPC they run in.

trainingJob_enableInterContainerTrafficEncryption :: Lens' TrainingJob (Maybe Bool)

To encrypt all communications between ML compute instances in distributed training, choose True. Encryption provides greater security for distributed training, but training might take longer. How long it takes depends on the amount of communication between compute instances, especially if you use a deep learning algorithm in distributed training.

trainingJob_trainingJobName :: Lens' TrainingJob (Maybe Text)

The name of the training job.

trainingJob_checkpointConfig :: Lens' TrainingJob (Maybe CheckpointConfig)

Undocumented member.

trainingJob_tuningJobArn :: Lens' TrainingJob (Maybe Text)

The Amazon Resource Name (ARN) of the associated hyperparameter tuning job if the training job was launched by a hyperparameter tuning job.

trainingJob_outputDataConfig :: Lens' TrainingJob (Maybe OutputDataConfig)

The S3 path where model artifacts that you configured when creating the job are stored. Amazon SageMaker creates subfolders for model artifacts.

trainingJob_environment :: Lens' TrainingJob (Maybe (HashMap Text Text))

The environment variables to set in the Docker container.

trainingJob_modelArtifacts :: Lens' TrainingJob (Maybe ModelArtifacts)

Information about the Amazon S3 location that is configured for storing model artifacts.

trainingJob_finalMetricDataList :: Lens' TrainingJob (Maybe [MetricData])

A list of final metric values that are set when the training job completes. Used only if the training job was configured to use metrics.

trainingJob_secondaryStatusTransitions :: Lens' TrainingJob (Maybe [SecondaryStatusTransition])

A history of all of the secondary statuses that the training job has transitioned through.

trainingJob_autoMLJobArn :: Lens' TrainingJob (Maybe Text)

The Amazon Resource Name (ARN) of the job.

trainingJob_failureReason :: Lens' TrainingJob (Maybe Text)

If the training job failed, the reason it failed.

trainingJob_tags :: Lens' TrainingJob (Maybe [Tag])

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

trainingJob_lastModifiedTime :: Lens' TrainingJob (Maybe UTCTime)

A timestamp that indicates when the status of the training job was last modified.

trainingJob_secondaryStatus :: Lens' TrainingJob (Maybe SecondaryStatus)

Provides detailed information about the state of the training job. For detailed information about the secondary status of the training job, see StatusMessage under SecondaryStatusTransition.

Amazon SageMaker provides primary statuses and secondary statuses that apply to each of them:

InProgress

- Starting - Starting the training job.

Downloading - An optional stage for algorithms that support File training input mode. It indicates that data is being downloaded to the ML storage volumes.

Training - Training is in progress.

Uploading - Training is complete and the model artifacts are being uploaded to the S3 location.

Completed

- Completed - The training job has completed.

Failed

- Failed - The training job has failed. The reason for the failure is returned in the FailureReason field of DescribeTrainingJobResponse.

Stopped

- MaxRuntimeExceeded - The job stopped because it exceeded the maximum allowed runtime.

Stopped - The training job has stopped.

Stopping

- Stopping - Stopping the training job.

Valid values for SecondaryStatus are subject to change.

We no longer support the following secondary statuses:

LaunchingMLInstances

PreparingTrainingStack

DownloadingTrainingImage

trainingJob_tensorBoardOutputConfig :: Lens' TrainingJob (Maybe TensorBoardOutputConfig)

Undocumented member.

trainingJob_debugRuleEvaluationStatuses :: Lens' TrainingJob (Maybe [DebugRuleEvaluationStatus])

Information about the evaluation status of the rules for the training job.

trainingJob_stoppingCondition :: Lens' TrainingJob (Maybe StoppingCondition)

Specifies a limit to how long a model training job can run. It also specifies how long a managed Spot training job has to complete. When the job reaches the time limit, Amazon SageMaker ends the training job. Use this API to cap model training costs.

To stop a job, Amazon SageMaker sends the algorithm the SIGTERM signal, which delays job termination for 120 seconds. Algorithms can use this 120-second window to save the model artifacts, so the results of training are not lost.

trainingJob_trainingJobStatus :: Lens' TrainingJob (Maybe TrainingJobStatus)

The status of the training job.

Training job statuses are:

InProgress - The training is in progress.

Completed - The training job has completed.

Failed - The training job has failed. To see the reason for the failure, see the FailureReason field in the response to a DescribeTrainingJobResponse call.

Stopping - The training job is stopping.

Stopped - The training job has stopped.

For more detailed information, see SecondaryStatus.

trainingJob_retryStrategy :: Lens' TrainingJob (Maybe RetryStrategy)

The number of times to retry the job when the job fails due to an InternalServerError.

trainingJob_resourceConfig :: Lens' TrainingJob (Maybe ResourceConfig)

Resources, including ML compute instances and ML storage volumes, that are configured for model training.

trainingJob_debugHookConfig :: Lens' TrainingJob (Maybe DebugHookConfig)

Undocumented member.

trainingJob_billableTimeInSeconds :: Lens' TrainingJob (Maybe Natural)

The billable time in seconds.

trainingJob_trainingStartTime :: Lens' TrainingJob (Maybe UTCTime)

Indicates the time when the training job starts on training instances. You are billed for the time interval between this time and the value of TrainingEndTime. The start time in CloudWatch Logs might be later than this time. The difference is due to the time it takes to download the training data and to the size of the training container.

trainingJob_trainingJobArn :: Lens' TrainingJob (Maybe Text)

The Amazon Resource Name (ARN) of the training job.

trainingJob_trainingEndTime :: Lens' TrainingJob (Maybe UTCTime)

Indicates the time when the training job ends on training instances. You are billed for the time interval between the value of TrainingStartTime and this time. For successful jobs and stopped jobs, this is the time after model artifacts are uploaded. For failed jobs, this is the time when Amazon SageMaker detects a job failure.

trainingJob_algorithmSpecification :: Lens' TrainingJob (Maybe AlgorithmSpecification)

Information about the algorithm used for training, and algorithm metadata.