Network.AWS.SageMaker.Types.TrainingJobDefinition

Description

Documentation

data TrainingJobDefinition

Defines the input needed to run a training job using the algorithm.

See: newTrainingJobDefinition smart constructor.

Constructors

TrainingJobDefinition'

Fields

hyperParameters :: Maybe (HashMap Text Text)

The hyperparameters used for the training job.

trainingInputMode :: TrainingInputMode

The input mode used by the algorithm for the training job. For the input modes that Amazon SageMaker algorithms support, see Algorithms.

If an algorithm supports the File input mode, Amazon SageMaker downloads the training data from S3 to the provisioned ML storage Volume, and mounts the directory to docker volume for training container. If an algorithm supports the Pipe input mode, Amazon SageMaker streams data directly from S3 to the container.

inputDataConfig :: NonEmpty Channel

An array of Channel objects, each of which specifies an input source.

outputDataConfig :: OutputDataConfig

the path to the S3 bucket where you want to store model artifacts. Amazon SageMaker creates subfolders for the artifacts.

resourceConfig :: ResourceConfig

The resources, including the ML compute instances and ML storage volumes, to use for model training.

stoppingCondition :: StoppingCondition

Specifies a limit to how long a model training job can run. It also specifies how long a managed Spot training job has to complete. When the job reaches the time limit, Amazon SageMaker ends the training job. Use this API to cap model training costs.

To stop a job, Amazon SageMaker sends the algorithm the SIGTERM signal, which delays job termination for 120 seconds. Algorithms can use this 120-second window to save the model artifacts.

newTrainingJobDefinition

Arguments

:: TrainingInputMode
$sel:trainingInputMode:TrainingJobDefinition'

-> NonEmpty Channel
$sel:inputDataConfig:TrainingJobDefinition'

-> OutputDataConfig
$sel:outputDataConfig:TrainingJobDefinition'

-> ResourceConfig
$sel:resourceConfig:TrainingJobDefinition'

-> StoppingCondition
$sel:stoppingCondition:TrainingJobDefinition'

-> TrainingJobDefinition

Create a value of TrainingJobDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hyperParameters:TrainingJobDefinition', trainingJobDefinition_hyperParameters - The hyperparameters used for the training job.

$sel:trainingInputMode:TrainingJobDefinition', trainingJobDefinition_trainingInputMode - The input mode used by the algorithm for the training job. For the input modes that Amazon SageMaker algorithms support, see Algorithms.

If an algorithm supports the File input mode, Amazon SageMaker downloads the training data from S3 to the provisioned ML storage Volume, and mounts the directory to docker volume for training container. If an algorithm supports the Pipe input mode, Amazon SageMaker streams data directly from S3 to the container.

$sel:inputDataConfig:TrainingJobDefinition', trainingJobDefinition_inputDataConfig - An array of Channel objects, each of which specifies an input source.

$sel:outputDataConfig:TrainingJobDefinition', trainingJobDefinition_outputDataConfig - the path to the S3 bucket where you want to store model artifacts. Amazon SageMaker creates subfolders for the artifacts.

$sel:resourceConfig:TrainingJobDefinition', trainingJobDefinition_resourceConfig - The resources, including the ML compute instances and ML storage volumes, to use for model training.

$sel:stoppingCondition:TrainingJobDefinition', trainingJobDefinition_stoppingCondition - Specifies a limit to how long a model training job can run. It also specifies how long a managed Spot training job has to complete. When the job reaches the time limit, Amazon SageMaker ends the training job. Use this API to cap model training costs.

To stop a job, Amazon SageMaker sends the algorithm the SIGTERM signal, which delays job termination for 120 seconds. Algorithms can use this 120-second window to save the model artifacts.

trainingJobDefinition_hyperParameters :: Lens' TrainingJobDefinition (Maybe (HashMap Text Text))

The hyperparameters used for the training job.

trainingJobDefinition_trainingInputMode :: Lens' TrainingJobDefinition TrainingInputMode

The input mode used by the algorithm for the training job. For the input modes that Amazon SageMaker algorithms support, see Algorithms.

If an algorithm supports the File input mode, Amazon SageMaker downloads the training data from S3 to the provisioned ML storage Volume, and mounts the directory to docker volume for training container. If an algorithm supports the Pipe input mode, Amazon SageMaker streams data directly from S3 to the container.

trainingJobDefinition_inputDataConfig :: Lens' TrainingJobDefinition (NonEmpty Channel)

An array of Channel objects, each of which specifies an input source.

trainingJobDefinition_outputDataConfig :: Lens' TrainingJobDefinition OutputDataConfig

the path to the S3 bucket where you want to store model artifacts. Amazon SageMaker creates subfolders for the artifacts.

trainingJobDefinition_resourceConfig :: Lens' TrainingJobDefinition ResourceConfig

The resources, including the ML compute instances and ML storage volumes, to use for model training.

trainingJobDefinition_stoppingCondition :: Lens' TrainingJobDefinition StoppingCondition

Specifies a limit to how long a model training job can run. It also specifies how long a managed Spot training job has to complete. When the job reaches the time limit, Amazon SageMaker ends the training job. Use this API to cap model training costs.

To stop a job, Amazon SageMaker sends the algorithm the SIGTERM signal, which delays job termination for 120 seconds. Algorithms can use this 120-second window to save the model artifacts.