Network.AWS.SageMaker.Types.TrainingJobStatusCounters

Description

Documentation

data TrainingJobStatusCounters

The numbers of training jobs launched by a hyperparameter tuning job, categorized by status.

See: newTrainingJobStatusCounters smart constructor.

Constructors

TrainingJobStatusCounters'

Fields

stopped :: Maybe Natural

The number of training jobs launched by a hyperparameter tuning job that were manually stopped.

retryableError :: Maybe Natural

The number of training jobs that failed, but can be retried. A failed training job can be retried only if it failed because an internal service error occurred.

inProgress :: Maybe Natural

The number of in-progress training jobs launched by a hyperparameter tuning job.

nonRetryableError :: Maybe Natural

The number of training jobs that failed and can't be retried. A failed training job can't be retried if it failed because a client error occurred.

completed :: Maybe Natural

The number of completed training jobs launched by the hyperparameter tuning job.

newTrainingJobStatusCounters :: TrainingJobStatusCounters

Create a value of TrainingJobStatusCounters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stopped:TrainingJobStatusCounters', trainingJobStatusCounters_stopped - The number of training jobs launched by a hyperparameter tuning job that were manually stopped.

$sel:retryableError:TrainingJobStatusCounters', trainingJobStatusCounters_retryableError - The number of training jobs that failed, but can be retried. A failed training job can be retried only if it failed because an internal service error occurred.

$sel:inProgress:TrainingJobStatusCounters', trainingJobStatusCounters_inProgress - The number of in-progress training jobs launched by a hyperparameter tuning job.

$sel:nonRetryableError:TrainingJobStatusCounters', trainingJobStatusCounters_nonRetryableError - The number of training jobs that failed and can't be retried. A failed training job can't be retried if it failed because a client error occurred.

$sel:completed:TrainingJobStatusCounters', trainingJobStatusCounters_completed - The number of completed training jobs launched by the hyperparameter tuning job.

trainingJobStatusCounters_stopped :: Lens' TrainingJobStatusCounters (Maybe Natural)

The number of training jobs launched by a hyperparameter tuning job that were manually stopped.

trainingJobStatusCounters_retryableError :: Lens' TrainingJobStatusCounters (Maybe Natural)

The number of training jobs that failed, but can be retried. A failed training job can be retried only if it failed because an internal service error occurred.

trainingJobStatusCounters_inProgress :: Lens' TrainingJobStatusCounters (Maybe Natural)

The number of in-progress training jobs launched by a hyperparameter tuning job.

trainingJobStatusCounters_nonRetryableError :: Lens' TrainingJobStatusCounters (Maybe Natural)

The number of training jobs that failed and can't be retried. A failed training job can't be retried if it failed because a client error occurred.

trainingJobStatusCounters_completed :: Lens' TrainingJobStatusCounters (Maybe Natural)

The number of completed training jobs launched by the hyperparameter tuning job.