Network.AWS.SageMaker.Types.TrainingSpecification

Description

Documentation

data TrainingSpecification

Defines how the algorithm is used for a training job.

See: newTrainingSpecification smart constructor.

Constructors

TrainingSpecification'

Fields

metricDefinitions :: Maybe [MetricDefinition]

A list of MetricDefinition objects, which are used for parsing metrics generated by the algorithm.

supportedHyperParameters :: Maybe [HyperParameterSpecification]

A list of the HyperParameterSpecification objects, that define the supported hyperparameters. This is required if the algorithm supports automatic model tuning.>

supportsDistributedTraining :: Maybe Bool

Indicates whether the algorithm supports distributed training. If set to false, buyers can't request more than one instance during training.

trainingImageDigest :: Maybe Text

An MD5 hash of the training algorithm that identifies the Docker image used for training.

supportedTuningJobObjectiveMetrics :: Maybe [HyperParameterTuningJobObjective]

A list of the metrics that the algorithm emits that can be used as the objective metric in a hyperparameter tuning job.

trainingImage :: Text

The Amazon ECR registry path of the Docker image that contains the training algorithm.

supportedTrainingInstanceTypes :: [TrainingInstanceType]

A list of the instance types that this algorithm can use for training.

trainingChannels :: NonEmpty ChannelSpecification

A list of ChannelSpecification objects, which specify the input sources to be used by the algorithm.

newTrainingSpecification

Arguments

:: Text
$sel:trainingImage:TrainingSpecification'

-> NonEmpty ChannelSpecification
$sel:trainingChannels:TrainingSpecification'

-> TrainingSpecification

Create a value of TrainingSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricDefinitions:TrainingSpecification', trainingSpecification_metricDefinitions - A list of MetricDefinition objects, which are used for parsing metrics generated by the algorithm.

$sel:supportedHyperParameters:TrainingSpecification', trainingSpecification_supportedHyperParameters - A list of the HyperParameterSpecification objects, that define the supported hyperparameters. This is required if the algorithm supports automatic model tuning.>

$sel:supportsDistributedTraining:TrainingSpecification', trainingSpecification_supportsDistributedTraining - Indicates whether the algorithm supports distributed training. If set to false, buyers can't request more than one instance during training.

$sel:trainingImageDigest:TrainingSpecification', trainingSpecification_trainingImageDigest - An MD5 hash of the training algorithm that identifies the Docker image used for training.

$sel:supportedTuningJobObjectiveMetrics:TrainingSpecification', trainingSpecification_supportedTuningJobObjectiveMetrics - A list of the metrics that the algorithm emits that can be used as the objective metric in a hyperparameter tuning job.

$sel:trainingImage:TrainingSpecification', trainingSpecification_trainingImage - The Amazon ECR registry path of the Docker image that contains the training algorithm.

$sel:supportedTrainingInstanceTypes:TrainingSpecification', trainingSpecification_supportedTrainingInstanceTypes - A list of the instance types that this algorithm can use for training.

$sel:trainingChannels:TrainingSpecification', trainingSpecification_trainingChannels - A list of ChannelSpecification objects, which specify the input sources to be used by the algorithm.

trainingSpecification_metricDefinitions :: Lens' TrainingSpecification (Maybe [MetricDefinition])

A list of MetricDefinition objects, which are used for parsing metrics generated by the algorithm.

trainingSpecification_supportedHyperParameters :: Lens' TrainingSpecification (Maybe [HyperParameterSpecification])

A list of the HyperParameterSpecification objects, that define the supported hyperparameters. This is required if the algorithm supports automatic model tuning.>

trainingSpecification_supportsDistributedTraining :: Lens' TrainingSpecification (Maybe Bool)

Indicates whether the algorithm supports distributed training. If set to false, buyers can't request more than one instance during training.

trainingSpecification_trainingImageDigest :: Lens' TrainingSpecification (Maybe Text)

An MD5 hash of the training algorithm that identifies the Docker image used for training.

trainingSpecification_supportedTuningJobObjectiveMetrics :: Lens' TrainingSpecification (Maybe [HyperParameterTuningJobObjective])

A list of the metrics that the algorithm emits that can be used as the objective metric in a hyperparameter tuning job.

trainingSpecification_trainingImage :: Lens' TrainingSpecification Text

The Amazon ECR registry path of the Docker image that contains the training algorithm.

trainingSpecification_supportedTrainingInstanceTypes :: Lens' TrainingSpecification [TrainingInstanceType]

A list of the instance types that this algorithm can use for training.

trainingSpecification_trainingChannels :: Lens' TrainingSpecification (NonEmpty ChannelSpecification)

A list of ChannelSpecification objects, which specify the input sources to be used by the algorithm.