Network.AWS.SageMaker.Types.TransformJob

Description

Documentation

data TransformJob

A batch transform job. For information about SageMaker batch transform, see Use Batch Transform.

See: newTransformJob smart constructor.

Constructors

TransformJob'

Fields

creationTime :: Maybe POSIX

A timestamp that shows when the transform Job was created.

labelingJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the labeling job that created the transform job.

transformJobName :: Maybe Text

The name of the transform job.

failureReason :: Maybe Text

If the transform job failed, the reason it failed.

modelClientConfig :: Maybe ModelClientConfig

batchStrategy :: Maybe BatchStrategy

Specifies the number of records to include in a mini-batch for an HTTP inference request. A record is a single unit of input data that inference can be made on. For example, a single line in a CSV file is a record.

maxPayloadInMB :: Maybe Natural

The maximum allowed size of the payload, in MB. A payload is the data portion of a record (without metadata). The value in MaxPayloadInMB must be greater than, or equal to, the size of a single record. To estimate the size of a record in MB, divide the size of your dataset by the number of records. To ensure that the records fit within the maximum payload size, we recommend using a slightly larger value. The default value is 6 MB. For cases where the payload might be arbitrarily large and is transmitted using HTTP chunked encoding, set the value to 0. This feature works only in supported algorithms. Currently, SageMaker built-in algorithms do not support HTTP chunked encoding.

environment :: Maybe (HashMap Text Text)

The environment variables to set in the Docker container. We support up to 16 key and values entries in the map.

transformResources :: Maybe TransformResources

modelName :: Maybe Text

The name of the model associated with the transform job.

experimentConfig :: Maybe ExperimentConfig

transformEndTime :: Maybe POSIX

Indicates when the transform job has been completed, or has stopped or failed. You are billed for the time interval between this time and the value of TransformStartTime.

transformStartTime :: Maybe POSIX

Indicates when the transform job starts on ML instances. You are billed for the time interval between this time and the value of TransformEndTime.

autoMLJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the AutoML job that created the transform job.

transformJobStatus :: Maybe TransformJobStatus

The status of the transform job.

Transform job statuses are:

InProgress - The job is in progress.

Completed - The job has completed.

Failed - The transform job has failed. To see the reason for the failure, see the FailureReason field in the response to a DescribeTransformJob call.

Stopping - The transform job is stopping.

Stopped - The transform job has stopped.

transformInput :: Maybe TransformInput

maxConcurrentTransforms :: Maybe Natural

The maximum number of parallel requests that can be sent to each instance in a transform job. If MaxConcurrentTransforms is set to 0 or left unset, SageMaker checks the optional execution-parameters to determine the settings for your chosen algorithm. If the execution-parameters endpoint is not enabled, the default value is 1. For built-in algorithms, you don't need to set a value for MaxConcurrentTransforms.

transformOutput :: Maybe TransformOutput

dataProcessing :: Maybe DataProcessing

transformJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the transform job.

tags :: Maybe [Tag]

A list of tags associated with the transform job.

newTransformJob :: TransformJob

Create a value of TransformJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:TransformJob', transformJob_creationTime - A timestamp that shows when the transform Job was created.

$sel:labelingJobArn:TransformJob', transformJob_labelingJobArn - The Amazon Resource Name (ARN) of the labeling job that created the transform job.

$sel:transformJobName:TransformJob', transformJob_transformJobName - The name of the transform job.

$sel:failureReason:TransformJob', transformJob_failureReason - If the transform job failed, the reason it failed.

$sel:modelClientConfig:TransformJob', transformJob_modelClientConfig - Undocumented member.

$sel:batchStrategy:TransformJob', transformJob_batchStrategy - Specifies the number of records to include in a mini-batch for an HTTP inference request. A record is a single unit of input data that inference can be made on. For example, a single line in a CSV file is a record.

$sel:maxPayloadInMB:TransformJob', transformJob_maxPayloadInMB - The maximum allowed size of the payload, in MB. A payload is the data portion of a record (without metadata). The value in MaxPayloadInMB must be greater than, or equal to, the size of a single record. To estimate the size of a record in MB, divide the size of your dataset by the number of records. To ensure that the records fit within the maximum payload size, we recommend using a slightly larger value. The default value is 6 MB. For cases where the payload might be arbitrarily large and is transmitted using HTTP chunked encoding, set the value to 0. This feature works only in supported algorithms. Currently, SageMaker built-in algorithms do not support HTTP chunked encoding.

$sel:environment:TransformJob', transformJob_environment - The environment variables to set in the Docker container. We support up to 16 key and values entries in the map.

$sel:transformResources:TransformJob', transformJob_transformResources - Undocumented member.

$sel:modelName:TransformJob', transformJob_modelName - The name of the model associated with the transform job.

$sel:experimentConfig:TransformJob', transformJob_experimentConfig - Undocumented member.

$sel:transformEndTime:TransformJob', transformJob_transformEndTime - Indicates when the transform job has been completed, or has stopped or failed. You are billed for the time interval between this time and the value of TransformStartTime.

$sel:transformStartTime:TransformJob', transformJob_transformStartTime - Indicates when the transform job starts on ML instances. You are billed for the time interval between this time and the value of TransformEndTime.

$sel:autoMLJobArn:TransformJob', transformJob_autoMLJobArn - The Amazon Resource Name (ARN) of the AutoML job that created the transform job.

$sel:transformJobStatus:TransformJob', transformJob_transformJobStatus - The status of the transform job.

Transform job statuses are:

InProgress - The job is in progress.

Completed - The job has completed.

Failed - The transform job has failed. To see the reason for the failure, see the FailureReason field in the response to a DescribeTransformJob call.

Stopping - The transform job is stopping.

Stopped - The transform job has stopped.

$sel:transformInput:TransformJob', transformJob_transformInput - Undocumented member.

$sel:maxConcurrentTransforms:TransformJob', transformJob_maxConcurrentTransforms - The maximum number of parallel requests that can be sent to each instance in a transform job. If MaxConcurrentTransforms is set to 0 or left unset, SageMaker checks the optional execution-parameters to determine the settings for your chosen algorithm. If the execution-parameters endpoint is not enabled, the default value is 1. For built-in algorithms, you don't need to set a value for MaxConcurrentTransforms.

$sel:transformOutput:TransformJob', transformJob_transformOutput - Undocumented member.

$sel:dataProcessing:TransformJob', transformJob_dataProcessing - Undocumented member.

$sel:transformJobArn:TransformJob', transformJob_transformJobArn - The Amazon Resource Name (ARN) of the transform job.

$sel:tags:TransformJob', transformJob_tags - A list of tags associated with the transform job.

transformJob_creationTime :: Lens' TransformJob (Maybe UTCTime)

A timestamp that shows when the transform Job was created.

transformJob_labelingJobArn :: Lens' TransformJob (Maybe Text)

The Amazon Resource Name (ARN) of the labeling job that created the transform job.

transformJob_transformJobName :: Lens' TransformJob (Maybe Text)

The name of the transform job.

transformJob_failureReason :: Lens' TransformJob (Maybe Text)

If the transform job failed, the reason it failed.

transformJob_modelClientConfig :: Lens' TransformJob (Maybe ModelClientConfig)

Undocumented member.

transformJob_batchStrategy :: Lens' TransformJob (Maybe BatchStrategy)

Specifies the number of records to include in a mini-batch for an HTTP inference request. A record is a single unit of input data that inference can be made on. For example, a single line in a CSV file is a record.

transformJob_maxPayloadInMB :: Lens' TransformJob (Maybe Natural)

The maximum allowed size of the payload, in MB. A payload is the data portion of a record (without metadata). The value in MaxPayloadInMB must be greater than, or equal to, the size of a single record. To estimate the size of a record in MB, divide the size of your dataset by the number of records. To ensure that the records fit within the maximum payload size, we recommend using a slightly larger value. The default value is 6 MB. For cases where the payload might be arbitrarily large and is transmitted using HTTP chunked encoding, set the value to 0. This feature works only in supported algorithms. Currently, SageMaker built-in algorithms do not support HTTP chunked encoding.

transformJob_environment :: Lens' TransformJob (Maybe (HashMap Text Text))

The environment variables to set in the Docker container. We support up to 16 key and values entries in the map.

transformJob_transformResources :: Lens' TransformJob (Maybe TransformResources)

Undocumented member.

transformJob_modelName :: Lens' TransformJob (Maybe Text)

The name of the model associated with the transform job.

transformJob_experimentConfig :: Lens' TransformJob (Maybe ExperimentConfig)

Undocumented member.

transformJob_transformEndTime :: Lens' TransformJob (Maybe UTCTime)

Indicates when the transform job has been completed, or has stopped or failed. You are billed for the time interval between this time and the value of TransformStartTime.

transformJob_transformStartTime :: Lens' TransformJob (Maybe UTCTime)

Indicates when the transform job starts on ML instances. You are billed for the time interval between this time and the value of TransformEndTime.

transformJob_autoMLJobArn :: Lens' TransformJob (Maybe Text)

The Amazon Resource Name (ARN) of the AutoML job that created the transform job.

transformJob_transformJobStatus :: Lens' TransformJob (Maybe TransformJobStatus)

The status of the transform job.

Transform job statuses are:

InProgress - The job is in progress.

Completed - The job has completed.

Failed - The transform job has failed. To see the reason for the failure, see the FailureReason field in the response to a DescribeTransformJob call.

Stopping - The transform job is stopping.

Stopped - The transform job has stopped.

transformJob_transformInput :: Lens' TransformJob (Maybe TransformInput)

Undocumented member.

transformJob_maxConcurrentTransforms :: Lens' TransformJob (Maybe Natural)

The maximum number of parallel requests that can be sent to each instance in a transform job. If MaxConcurrentTransforms is set to 0 or left unset, SageMaker checks the optional execution-parameters to determine the settings for your chosen algorithm. If the execution-parameters endpoint is not enabled, the default value is 1. For built-in algorithms, you don't need to set a value for MaxConcurrentTransforms.

transformJob_transformOutput :: Lens' TransformJob (Maybe TransformOutput)

Undocumented member.

transformJob_dataProcessing :: Lens' TransformJob (Maybe DataProcessing)

Undocumented member.

transformJob_transformJobArn :: Lens' TransformJob (Maybe Text)

The Amazon Resource Name (ARN) of the transform job.

transformJob_tags :: Lens' TransformJob (Maybe [Tag])

A list of tags associated with the transform job.