Network.AWS.SageMaker.Types.TransformJobDefinition

Description

Documentation

data TransformJobDefinition

Defines the input needed to run a transform job using the inference specification specified in the algorithm.

See: newTransformJobDefinition smart constructor.

Constructors

TransformJobDefinition'

Fields

maxConcurrentTransforms :: Maybe Natural

The maximum number of parallel requests that can be sent to each instance in a transform job. The default value is 1.

environment :: Maybe (HashMap Text Text)

The environment variables to set in the Docker container. We support up to 16 key and values entries in the map.

maxPayloadInMB :: Maybe Natural

The maximum payload size allowed, in MB. A payload is the data portion of a record (without metadata).

batchStrategy :: Maybe BatchStrategy

A string that determines the number of records included in a single mini-batch.

SingleRecord means only one record is used per mini-batch. MultiRecord means a mini-batch is set to contain as many records that can fit within the MaxPayloadInMB limit.

transformInput :: TransformInput

A description of the input source and the way the transform job consumes it.

transformOutput :: TransformOutput

Identifies the Amazon S3 location where you want Amazon SageMaker to save the results from the transform job.

transformResources :: TransformResources

Identifies the ML compute instances for the transform job.

newTransformJobDefinition

Arguments

:: TransformInput
$sel:transformInput:TransformJobDefinition'

-> TransformOutput
$sel:transformOutput:TransformJobDefinition'

-> TransformResources
$sel:transformResources:TransformJobDefinition'

-> TransformJobDefinition

Create a value of TransformJobDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxConcurrentTransforms:TransformJobDefinition', transformJobDefinition_maxConcurrentTransforms - The maximum number of parallel requests that can be sent to each instance in a transform job. The default value is 1.

$sel:environment:TransformJobDefinition', transformJobDefinition_environment - The environment variables to set in the Docker container. We support up to 16 key and values entries in the map.

$sel:maxPayloadInMB:TransformJobDefinition', transformJobDefinition_maxPayloadInMB - The maximum payload size allowed, in MB. A payload is the data portion of a record (without metadata).

$sel:batchStrategy:TransformJobDefinition', transformJobDefinition_batchStrategy - A string that determines the number of records included in a single mini-batch.

SingleRecord means only one record is used per mini-batch. MultiRecord means a mini-batch is set to contain as many records that can fit within the MaxPayloadInMB limit.

$sel:transformInput:TransformJobDefinition', transformJobDefinition_transformInput - A description of the input source and the way the transform job consumes it.

$sel:transformOutput:TransformJobDefinition', transformJobDefinition_transformOutput - Identifies the Amazon S3 location where you want Amazon SageMaker to save the results from the transform job.

$sel:transformResources:TransformJobDefinition', transformJobDefinition_transformResources - Identifies the ML compute instances for the transform job.

transformJobDefinition_maxConcurrentTransforms :: Lens' TransformJobDefinition (Maybe Natural)

The maximum number of parallel requests that can be sent to each instance in a transform job. The default value is 1.

transformJobDefinition_environment :: Lens' TransformJobDefinition (Maybe (HashMap Text Text))

The environment variables to set in the Docker container. We support up to 16 key and values entries in the map.

transformJobDefinition_maxPayloadInMB :: Lens' TransformJobDefinition (Maybe Natural)

The maximum payload size allowed, in MB. A payload is the data portion of a record (without metadata).

transformJobDefinition_batchStrategy :: Lens' TransformJobDefinition (Maybe BatchStrategy)

A string that determines the number of records included in a single mini-batch.

SingleRecord means only one record is used per mini-batch. MultiRecord means a mini-batch is set to contain as many records that can fit within the MaxPayloadInMB limit.

transformJobDefinition_transformInput :: Lens' TransformJobDefinition TransformInput

A description of the input source and the way the transform job consumes it.

transformJobDefinition_transformOutput :: Lens' TransformJobDefinition TransformOutput

Identifies the Amazon S3 location where you want Amazon SageMaker to save the results from the transform job.

transformJobDefinition_transformResources :: Lens' TransformJobDefinition TransformResources

Identifies the ML compute instances for the transform job.