Network.AWS.SageMaker.Types.TransformJobSummary

Description

Documentation

data TransformJobSummary

Provides a summary of a transform job. Multiple TransformJobSummary objects are returned as a list after in response to a ListTransformJobs call.

See: newTransformJobSummary smart constructor.

Constructors

TransformJobSummary'

Fields

failureReason :: Maybe Text

If the transform job failed, the reason it failed.

lastModifiedTime :: Maybe POSIX

Indicates when the transform job was last modified.

transformEndTime :: Maybe POSIX

Indicates when the transform job ends on compute instances. For successful jobs and stopped jobs, this is the exact time recorded after the results are uploaded. For failed jobs, this is when Amazon SageMaker detected that the job failed.

transformJobName :: Text

The name of the transform job.

transformJobArn :: Text

The Amazon Resource Name (ARN) of the transform job.

creationTime :: POSIX

A timestamp that shows when the transform Job was created.

transformJobStatus :: TransformJobStatus

The status of the transform job.

newTransformJobSummary

Arguments

:: Text
$sel:transformJobName:TransformJobSummary'

-> Text
$sel:transformJobArn:TransformJobSummary'

-> UTCTime
$sel:creationTime:TransformJobSummary'

-> TransformJobStatus
$sel:transformJobStatus:TransformJobSummary'

-> TransformJobSummary

Create a value of TransformJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:TransformJobSummary', transformJobSummary_failureReason - If the transform job failed, the reason it failed.

$sel:lastModifiedTime:TransformJobSummary', transformJobSummary_lastModifiedTime - Indicates when the transform job was last modified.

$sel:transformEndTime:TransformJobSummary', transformJobSummary_transformEndTime - Indicates when the transform job ends on compute instances. For successful jobs and stopped jobs, this is the exact time recorded after the results are uploaded. For failed jobs, this is when Amazon SageMaker detected that the job failed.

$sel:transformJobName:TransformJobSummary', transformJobSummary_transformJobName - The name of the transform job.

$sel:transformJobArn:TransformJobSummary', transformJobSummary_transformJobArn - The Amazon Resource Name (ARN) of the transform job.

$sel:creationTime:TransformJobSummary', transformJobSummary_creationTime - A timestamp that shows when the transform Job was created.

$sel:transformJobStatus:TransformJobSummary', transformJobSummary_transformJobStatus - The status of the transform job.

transformJobSummary_failureReason :: Lens' TransformJobSummary (Maybe Text)

If the transform job failed, the reason it failed.

transformJobSummary_lastModifiedTime :: Lens' TransformJobSummary (Maybe UTCTime)

Indicates when the transform job was last modified.

transformJobSummary_transformEndTime :: Lens' TransformJobSummary (Maybe UTCTime)

Indicates when the transform job ends on compute instances. For successful jobs and stopped jobs, this is the exact time recorded after the results are uploaded. For failed jobs, this is when Amazon SageMaker detected that the job failed.

transformJobSummary_transformJobName :: Lens' TransformJobSummary Text

The name of the transform job.

transformJobSummary_transformJobArn :: Lens' TransformJobSummary Text

The Amazon Resource Name (ARN) of the transform job.

transformJobSummary_creationTime :: Lens' TransformJobSummary UTCTime

A timestamp that shows when the transform Job was created.

transformJobSummary_transformJobStatus :: Lens' TransformJobSummary TransformJobStatus

The status of the transform job.