Network.AWS.SageMaker.Types.TransformS3DataSource

Description

Documentation

data TransformS3DataSource

Describes the S3 data source.

See: newTransformS3DataSource smart constructor.

Constructors

TransformS3DataSource'

Fields

s3DataType :: S3DataType

If you choose S3Prefix, S3Uri identifies a key name prefix. Amazon SageMaker uses all objects with the specified key name prefix for batch transform.

If you choose ManifestFile, S3Uri identifies an object that is a manifest file containing a list of object keys that you want Amazon SageMaker to use for batch transform.

The following values are compatible: ManifestFile, S3Prefix

The following value is not compatible: AugmentedManifestFile

s3Uri :: Text

Depending on the value specified for the S3DataType, identifies either a key name prefix or a manifest. For example:

A key name prefix might look like this: s3://bucketname/exampleprefix.

A manifest might look like this: s3://bucketname/example.manifest

The manifest is an S3 object which is a JSON file with the following format:

[ {"prefix": "s3://customer_bucket/some/prefix/"},

"relative/path/to/custdata-1",

"relative/path/custdata-2",

...

"relative/path/custdata-N"

]

The preceding JSON matches the following S3Uris:

s3://customer_bucket/some/prefix/relative/path/to/custdata-1

s3://customer_bucket/some/prefix/relative/path/custdata-2

...

s3://customer_bucket/some/prefix/relative/path/custdata-N

The complete set of S3Uris in this manifest constitutes the input data for the channel for this datasource. The object that each S3Uris points to must be readable by the IAM role that Amazon SageMaker uses to perform tasks on your behalf.

newTransformS3DataSource

Arguments

:: S3DataType
$sel:s3DataType:TransformS3DataSource'

-> Text
$sel:s3Uri:TransformS3DataSource'

-> TransformS3DataSource

Create a value of TransformS3DataSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3DataType:TransformS3DataSource', transformS3DataSource_s3DataType - If you choose S3Prefix, S3Uri identifies a key name prefix. Amazon SageMaker uses all objects with the specified key name prefix for batch transform.

If you choose ManifestFile, S3Uri identifies an object that is a manifest file containing a list of object keys that you want Amazon SageMaker to use for batch transform.

The following values are compatible: ManifestFile, S3Prefix

The following value is not compatible: AugmentedManifestFile

$sel:s3Uri:TransformS3DataSource', transformS3DataSource_s3Uri - Depending on the value specified for the S3DataType, identifies either a key name prefix or a manifest. For example:

A key name prefix might look like this: s3://bucketname/exampleprefix.

A manifest might look like this: s3://bucketname/example.manifest

The manifest is an S3 object which is a JSON file with the following format:

[ {"prefix": "s3://customer_bucket/some/prefix/"},

"relative/path/to/custdata-1",

"relative/path/custdata-2",

...

"relative/path/custdata-N"

]

The preceding JSON matches the following S3Uris:

s3://customer_bucket/some/prefix/relative/path/to/custdata-1

s3://customer_bucket/some/prefix/relative/path/custdata-2

...

s3://customer_bucket/some/prefix/relative/path/custdata-N

The complete set of S3Uris in this manifest constitutes the input data for the channel for this datasource. The object that each S3Uris points to must be readable by the IAM role that Amazon SageMaker uses to perform tasks on your behalf.

transformS3DataSource_s3DataType :: Lens' TransformS3DataSource S3DataType

If you choose S3Prefix, S3Uri identifies a key name prefix. Amazon SageMaker uses all objects with the specified key name prefix for batch transform.

If you choose ManifestFile, S3Uri identifies an object that is a manifest file containing a list of object keys that you want Amazon SageMaker to use for batch transform.

The following values are compatible: ManifestFile, S3Prefix

The following value is not compatible: AugmentedManifestFile

transformS3DataSource_s3Uri :: Lens' TransformS3DataSource Text

Depending on the value specified for the S3DataType, identifies either a key name prefix or a manifest. For example:

A key name prefix might look like this: s3://bucketname/exampleprefix.

A manifest might look like this: s3://bucketname/example.manifest

The manifest is an S3 object which is a JSON file with the following format:

[ {"prefix": "s3://customer_bucket/some/prefix/"},

"relative/path/to/custdata-1",

"relative/path/custdata-2",

...

"relative/path/custdata-N"

]

The preceding JSON matches the following S3Uris:

s3://customer_bucket/some/prefix/relative/path/to/custdata-1

s3://customer_bucket/some/prefix/relative/path/custdata-2

...

s3://customer_bucket/some/prefix/relative/path/custdata-N

The complete set of S3Uris in this manifest constitutes the input data for the channel for this datasource. The object that each S3Uris points to must be readable by the IAM role that Amazon SageMaker uses to perform tasks on your behalf.