Network.AWS.SageMaker.Types.Trial

Description

Documentation

data Trial

The properties of a trial as returned by the Search API.

See: newTrial smart constructor.

Constructors

Trial'

Fields

creationTime :: Maybe POSIX

When the trial was created.

metadataProperties :: Maybe MetadataProperties

trialComponentSummaries :: Maybe [TrialComponentSimpleSummary]

A list of the components associated with the trial. For each component, a summary of the component's properties is included.

trialArn :: Maybe Text

The Amazon Resource Name (ARN) of the trial.

createdBy :: Maybe UserContext

Who created the trial.

lastModifiedTime :: Maybe POSIX

Who last modified the trial.

experimentName :: Maybe Text

The name of the experiment the trial is part of.

source :: Maybe TrialSource

displayName :: Maybe Text

The name of the trial as displayed. If DisplayName isn't specified, TrialName is displayed.

trialName :: Maybe Text

The name of the trial.

lastModifiedBy :: Maybe UserContext

tags :: Maybe [Tag]

The list of tags that are associated with the trial. You can use Search API to search on the tags.

newTrial :: Trial

Create a value of Trial with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Trial', trial_creationTime - When the trial was created.

$sel:metadataProperties:Trial', trial_metadataProperties - Undocumented member.

$sel:trialComponentSummaries:Trial', trial_trialComponentSummaries - A list of the components associated with the trial. For each component, a summary of the component's properties is included.

$sel:trialArn:Trial', trial_trialArn - The Amazon Resource Name (ARN) of the trial.

$sel:createdBy:Trial', trial_createdBy - Who created the trial.

$sel:lastModifiedTime:Trial', trial_lastModifiedTime - Who last modified the trial.

$sel:experimentName:Trial', trial_experimentName - The name of the experiment the trial is part of.

$sel:source:Trial', trial_source - Undocumented member.

$sel:displayName:Trial', trial_displayName - The name of the trial as displayed. If DisplayName isn't specified, TrialName is displayed.

$sel:trialName:Trial', trial_trialName - The name of the trial.

$sel:lastModifiedBy:Trial', trial_lastModifiedBy - Undocumented member.

$sel:tags:Trial', trial_tags - The list of tags that are associated with the trial. You can use Search API to search on the tags.

trial_creationTime :: Lens' Trial (Maybe UTCTime)

When the trial was created.

trial_metadataProperties :: Lens' Trial (Maybe MetadataProperties)

Undocumented member.

trial_trialComponentSummaries :: Lens' Trial (Maybe [TrialComponentSimpleSummary])

A list of the components associated with the trial. For each component, a summary of the component's properties is included.

trial_trialArn :: Lens' Trial (Maybe Text)

The Amazon Resource Name (ARN) of the trial.

trial_createdBy :: Lens' Trial (Maybe UserContext)

Who created the trial.

trial_lastModifiedTime :: Lens' Trial (Maybe UTCTime)

Who last modified the trial.

trial_experimentName :: Lens' Trial (Maybe Text)

The name of the experiment the trial is part of.

trial_source :: Lens' Trial (Maybe TrialSource)

Undocumented member.

trial_displayName :: Lens' Trial (Maybe Text)

The name of the trial as displayed. If DisplayName isn't specified, TrialName is displayed.

trial_trialName :: Lens' Trial (Maybe Text)

The name of the trial.

trial_lastModifiedBy :: Lens' Trial (Maybe UserContext)

Undocumented member.

trial_tags :: Lens' Trial (Maybe [Tag])

The list of tags that are associated with the trial. You can use Search API to search on the tags.