Network.AWS.SageMaker.Types.TrialComponent

Description

Documentation

data TrialComponent

The properties of a trial component as returned by the Search API.

See: newTrialComponent smart constructor.

Constructors

TrialComponent'

Fields

parents :: Maybe [Parent]

An array of the parents of the component. A parent is a trial the component is associated with and the experiment the trial is part of. A component might not have any parents.

creationTime :: Maybe POSIX

When the component was created.

status :: Maybe TrialComponentStatus

metadataProperties :: Maybe MetadataProperties

sourceDetail :: Maybe TrialComponentSourceDetail

Details of the source of the component.

trialComponentArn :: Maybe Text

The Amazon Resource Name (ARN) of the trial component.

startTime :: Maybe POSIX

When the component started.

source :: Maybe TrialComponentSource

The Amazon Resource Name (ARN) and job type of the source of the component.

endTime :: Maybe POSIX

When the component ended.

metrics :: Maybe [TrialComponentMetricSummary]

The metrics for the component.

tags :: Maybe [Tag]

The list of tags that are associated with the component. You can use Search API to search on the tags.

lastModifiedTime :: Maybe POSIX

When the component was last modified.

createdBy :: Maybe UserContext

Who created the trial component.

inputArtifacts :: Maybe (HashMap Text TrialComponentArtifact)

The input artifacts of the component.

lastModifiedBy :: Maybe UserContext

displayName :: Maybe Text

The name of the component as displayed. If DisplayName isn't specified, TrialComponentName is displayed.

trialComponentName :: Maybe Text

The name of the trial component.

outputArtifacts :: Maybe (HashMap Text TrialComponentArtifact)

The output artifacts of the component.

parameters :: Maybe (HashMap Text TrialComponentParameterValue)

The hyperparameters of the component.

newTrialComponent :: TrialComponent

Create a value of TrialComponent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parents:TrialComponent', trialComponent_parents - An array of the parents of the component. A parent is a trial the component is associated with and the experiment the trial is part of. A component might not have any parents.

$sel:creationTime:TrialComponent', trialComponent_creationTime - When the component was created.

$sel:status:TrialComponent', trialComponent_status - Undocumented member.

$sel:metadataProperties:TrialComponent', trialComponent_metadataProperties - Undocumented member.

$sel:sourceDetail:TrialComponent', trialComponent_sourceDetail - Details of the source of the component.

$sel:trialComponentArn:TrialComponent', trialComponent_trialComponentArn - The Amazon Resource Name (ARN) of the trial component.

$sel:startTime:TrialComponent', trialComponent_startTime - When the component started.

$sel:source:TrialComponent', trialComponent_source - The Amazon Resource Name (ARN) and job type of the source of the component.

$sel:endTime:TrialComponent', trialComponent_endTime - When the component ended.

$sel:metrics:TrialComponent', trialComponent_metrics - The metrics for the component.

$sel:tags:TrialComponent', trialComponent_tags - The list of tags that are associated with the component. You can use Search API to search on the tags.

$sel:lastModifiedTime:TrialComponent', trialComponent_lastModifiedTime - When the component was last modified.

$sel:createdBy:TrialComponent', trialComponent_createdBy - Who created the trial component.

$sel:inputArtifacts:TrialComponent', trialComponent_inputArtifacts - The input artifacts of the component.

$sel:lastModifiedBy:TrialComponent', trialComponent_lastModifiedBy - Undocumented member.

$sel:displayName:TrialComponent', trialComponent_displayName - The name of the component as displayed. If DisplayName isn't specified, TrialComponentName is displayed.

$sel:trialComponentName:TrialComponent', trialComponent_trialComponentName - The name of the trial component.

$sel:outputArtifacts:TrialComponent', trialComponent_outputArtifacts - The output artifacts of the component.

$sel:parameters:TrialComponent', trialComponent_parameters - The hyperparameters of the component.

trialComponent_parents :: Lens' TrialComponent (Maybe [Parent])

An array of the parents of the component. A parent is a trial the component is associated with and the experiment the trial is part of. A component might not have any parents.

trialComponent_creationTime :: Lens' TrialComponent (Maybe UTCTime)

When the component was created.

trialComponent_status :: Lens' TrialComponent (Maybe TrialComponentStatus)

Undocumented member.

trialComponent_metadataProperties :: Lens' TrialComponent (Maybe MetadataProperties)

Undocumented member.

trialComponent_sourceDetail :: Lens' TrialComponent (Maybe TrialComponentSourceDetail)

Details of the source of the component.

trialComponent_trialComponentArn :: Lens' TrialComponent (Maybe Text)

The Amazon Resource Name (ARN) of the trial component.

trialComponent_startTime :: Lens' TrialComponent (Maybe UTCTime)

When the component started.

trialComponent_source :: Lens' TrialComponent (Maybe TrialComponentSource)

The Amazon Resource Name (ARN) and job type of the source of the component.

trialComponent_endTime :: Lens' TrialComponent (Maybe UTCTime)

When the component ended.

trialComponent_metrics :: Lens' TrialComponent (Maybe [TrialComponentMetricSummary])

The metrics for the component.

trialComponent_tags :: Lens' TrialComponent (Maybe [Tag])

The list of tags that are associated with the component. You can use Search API to search on the tags.

trialComponent_lastModifiedTime :: Lens' TrialComponent (Maybe UTCTime)

When the component was last modified.

trialComponent_createdBy :: Lens' TrialComponent (Maybe UserContext)

Who created the trial component.

trialComponent_inputArtifacts :: Lens' TrialComponent (Maybe (HashMap Text TrialComponentArtifact))

The input artifacts of the component.

trialComponent_lastModifiedBy :: Lens' TrialComponent (Maybe UserContext)

Undocumented member.

trialComponent_displayName :: Lens' TrialComponent (Maybe Text)

The name of the component as displayed. If DisplayName isn't specified, TrialComponentName is displayed.

trialComponent_trialComponentName :: Lens' TrialComponent (Maybe Text)

The name of the trial component.

trialComponent_outputArtifacts :: Lens' TrialComponent (Maybe (HashMap Text TrialComponentArtifact))

The output artifacts of the component.

trialComponent_parameters :: Lens' TrialComponent (Maybe (HashMap Text TrialComponentParameterValue))

The hyperparameters of the component.