Network.AWS.SageMaker.Types.TrialComponentMetricSummary

Description

Documentation

data TrialComponentMetricSummary

A summary of the metrics of a trial component.

See: newTrialComponentMetricSummary smart constructor.

Constructors

TrialComponentMetricSummary'

Fields

max :: Maybe Double

The maximum value of the metric.

sourceArn :: Maybe Text

The Amazon Resource Name (ARN) of the source.

avg :: Maybe Double

The average value of the metric.

count :: Maybe Int

The number of samples used to generate the metric.

metricName :: Maybe Text

The name of the metric.

stdDev :: Maybe Double

The standard deviation of the metric.

min :: Maybe Double

The minimum value of the metric.

last :: Maybe Double

The most recent value of the metric.

timeStamp :: Maybe POSIX

When the metric was last updated.

newTrialComponentMetricSummary :: TrialComponentMetricSummary

Create a value of TrialComponentMetricSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:max:TrialComponentMetricSummary', trialComponentMetricSummary_max - The maximum value of the metric.

$sel:sourceArn:TrialComponentMetricSummary', trialComponentMetricSummary_sourceArn - The Amazon Resource Name (ARN) of the source.

$sel:avg:TrialComponentMetricSummary', trialComponentMetricSummary_avg - The average value of the metric.

$sel:count:TrialComponentMetricSummary', trialComponentMetricSummary_count - The number of samples used to generate the metric.

$sel:metricName:TrialComponentMetricSummary', trialComponentMetricSummary_metricName - The name of the metric.

$sel:stdDev:TrialComponentMetricSummary', trialComponentMetricSummary_stdDev - The standard deviation of the metric.

$sel:min:TrialComponentMetricSummary', trialComponentMetricSummary_min - The minimum value of the metric.

$sel:last:TrialComponentMetricSummary', trialComponentMetricSummary_last - The most recent value of the metric.

$sel:timeStamp:TrialComponentMetricSummary', trialComponentMetricSummary_timeStamp - When the metric was last updated.

trialComponentMetricSummary_max :: Lens' TrialComponentMetricSummary (Maybe Double)

The maximum value of the metric.

trialComponentMetricSummary_sourceArn :: Lens' TrialComponentMetricSummary (Maybe Text)

The Amazon Resource Name (ARN) of the source.

trialComponentMetricSummary_avg :: Lens' TrialComponentMetricSummary (Maybe Double)

The average value of the metric.

trialComponentMetricSummary_count :: Lens' TrialComponentMetricSummary (Maybe Int)

The number of samples used to generate the metric.

trialComponentMetricSummary_metricName :: Lens' TrialComponentMetricSummary (Maybe Text)

The name of the metric.

trialComponentMetricSummary_stdDev :: Lens' TrialComponentMetricSummary (Maybe Double)

The standard deviation of the metric.

trialComponentMetricSummary_min :: Lens' TrialComponentMetricSummary (Maybe Double)

The minimum value of the metric.

trialComponentMetricSummary_last :: Lens' TrialComponentMetricSummary (Maybe Double)

The most recent value of the metric.

trialComponentMetricSummary_timeStamp :: Lens' TrialComponentMetricSummary (Maybe UTCTime)

When the metric was last updated.