Network.AWS.SageMaker.Types.TrialComponentSummary

Description

Documentation

data TrialComponentSummary

A summary of the properties of a trial component. To get all the properties, call the DescribeTrialComponent API and provide the TrialComponentName.

See: newTrialComponentSummary smart constructor.

Constructors

TrialComponentSummary'

Fields

creationTime :: Maybe POSIX

When the component was created.

status :: Maybe TrialComponentStatus

The status of the component. States include:

InProgress

Completed

Failed

startTime :: Maybe POSIX

When the component started.

createdBy :: Maybe UserContext

Who created the trial component.

lastModifiedTime :: Maybe POSIX

When the component was last modified.

endTime :: Maybe POSIX

When the component ended.

trialComponentName :: Maybe Text

The name of the trial component.

displayName :: Maybe Text

The name of the component as displayed. If DisplayName isn't specified, TrialComponentName is displayed.

lastModifiedBy :: Maybe UserContext

Who last modified the component.

trialComponentArn :: Maybe Text

The ARN of the trial component.

trialComponentSource :: Maybe TrialComponentSource

newTrialComponentSummary :: TrialComponentSummary

Create a value of TrialComponentSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:TrialComponentSummary', trialComponentSummary_creationTime - When the component was created.

$sel:status:TrialComponentSummary', trialComponentSummary_status - The status of the component. States include:

InProgress

Completed

Failed

$sel:startTime:TrialComponentSummary', trialComponentSummary_startTime - When the component started.

$sel:createdBy:TrialComponentSummary', trialComponentSummary_createdBy - Who created the trial component.

$sel:lastModifiedTime:TrialComponentSummary', trialComponentSummary_lastModifiedTime - When the component was last modified.

$sel:endTime:TrialComponentSummary', trialComponentSummary_endTime - When the component ended.

$sel:trialComponentName:TrialComponentSummary', trialComponentSummary_trialComponentName - The name of the trial component.

$sel:displayName:TrialComponentSummary', trialComponentSummary_displayName - The name of the component as displayed. If DisplayName isn't specified, TrialComponentName is displayed.

$sel:lastModifiedBy:TrialComponentSummary', trialComponentSummary_lastModifiedBy - Who last modified the component.

$sel:trialComponentArn:TrialComponentSummary', trialComponentSummary_trialComponentArn - The ARN of the trial component.

$sel:trialComponentSource:TrialComponentSummary', trialComponentSummary_trialComponentSource - Undocumented member.

trialComponentSummary_creationTime :: Lens' TrialComponentSummary (Maybe UTCTime)

When the component was created.

trialComponentSummary_status :: Lens' TrialComponentSummary (Maybe TrialComponentStatus)

The status of the component. States include:

InProgress

Completed

Failed

trialComponentSummary_startTime :: Lens' TrialComponentSummary (Maybe UTCTime)

When the component started.

trialComponentSummary_createdBy :: Lens' TrialComponentSummary (Maybe UserContext)

Who created the trial component.

trialComponentSummary_lastModifiedTime :: Lens' TrialComponentSummary (Maybe UTCTime)

When the component was last modified.

trialComponentSummary_endTime :: Lens' TrialComponentSummary (Maybe UTCTime)

When the component ended.

trialComponentSummary_trialComponentName :: Lens' TrialComponentSummary (Maybe Text)

The name of the trial component.

trialComponentSummary_displayName :: Lens' TrialComponentSummary (Maybe Text)

The name of the component as displayed. If DisplayName isn't specified, TrialComponentName is displayed.

trialComponentSummary_lastModifiedBy :: Lens' TrialComponentSummary (Maybe UserContext)

Who last modified the component.

trialComponentSummary_trialComponentArn :: Lens' TrialComponentSummary (Maybe Text)

The ARN of the trial component.

trialComponentSummary_trialComponentSource :: Lens' TrialComponentSummary (Maybe TrialComponentSource)

Undocumented member.