Network.AWS.SageMaker.Types.TuningJobStepMetaData

Description

Documentation

data TuningJobStepMetaData

Metadata for a tuning step.

See: newTuningJobStepMetaData smart constructor.

Constructors

TuningJobStepMetaData'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the tuning job that was run by this step execution.

newTuningJobStepMetaData :: TuningJobStepMetaData

Create a value of TuningJobStepMetaData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:TuningJobStepMetaData', tuningJobStepMetaData_arn - The Amazon Resource Name (ARN) of the tuning job that was run by this step execution.

tuningJobStepMetaData_arn :: Lens' TuningJobStepMetaData (Maybe Text)

The Amazon Resource Name (ARN) of the tuning job that was run by this step execution.