Network.AWS.SageMaker.Types.UiConfig

Description

Documentation

data UiConfig

Provided configuration information for the worker UI for a labeling job. Provide either HumanTaskUiArn or UiTemplateS3Uri.

For named entity recognition, 3D point cloud and video frame labeling jobs, use HumanTaskUiArn.

For all other Ground Truth built-in task types and custom task types, use UiTemplateS3Uri to specify the location of a worker task template in Amazon S3.

See: newUiConfig smart constructor.

Constructors

UiConfig'

Fields

uiTemplateS3Uri :: Maybe Text

The Amazon S3 bucket location of the UI template, or worker task template. This is the template used to render the worker UI and tools for labeling job tasks. For more information about the contents of a UI template, see Creating Your Custom Labeling Task Template.

humanTaskUiArn :: Maybe Text

The ARN of the worker task template used to render the worker UI and tools for labeling job tasks.

Use this parameter when you are creating a labeling job for named entity recognition, 3D point cloud and video frame labeling jobs. Use your labeling job task type to select one of the following ARNs and use it with this parameter when you create a labeling job. Replace aws-region with the Amazon Web Services Region you are creating your labeling job in. For example, replace aws-region with us-west-1 if you create a labeling job in US West (N. California).

Named Entity Recognition

Use the following HumanTaskUiArn for named entity recognition labeling jobs:

arn:aws:sagemaker:aws-region:394669845002:human-task-ui/NamedEntityRecognition

3D Point Cloud HumanTaskUiArns

Use this HumanTaskUiArn for 3D point cloud object detection and 3D point cloud object detection adjustment labeling jobs.

arn:aws:sagemaker:aws-region:394669845002:human-task-ui/PointCloudObjectDetection

Use this HumanTaskUiArn for 3D point cloud object tracking and 3D point cloud object tracking adjustment labeling jobs.

arn:aws:sagemaker:aws-region:394669845002:human-task-ui/PointCloudObjectTracking

Use this HumanTaskUiArn for 3D point cloud semantic segmentation and 3D point cloud semantic segmentation adjustment labeling jobs.

arn:aws:sagemaker:aws-region:394669845002:human-task-ui/PointCloudSemanticSegmentation

Video Frame HumanTaskUiArns

Use this HumanTaskUiArn for video frame object detection and video frame object detection adjustment labeling jobs.

arn:aws:sagemaker:region:394669845002:human-task-ui/VideoObjectDetection

Use this HumanTaskUiArn for video frame object tracking and video frame object tracking adjustment labeling jobs.

arn:aws:sagemaker:aws-region:394669845002:human-task-ui/VideoObjectTracking

newUiConfig :: UiConfig

Create a value of UiConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:uiTemplateS3Uri:UiConfig', uiConfig_uiTemplateS3Uri - The Amazon S3 bucket location of the UI template, or worker task template. This is the template used to render the worker UI and tools for labeling job tasks. For more information about the contents of a UI template, see Creating Your Custom Labeling Task Template.

$sel:humanTaskUiArn:UiConfig', uiConfig_humanTaskUiArn - The ARN of the worker task template used to render the worker UI and tools for labeling job tasks.

Use this parameter when you are creating a labeling job for named entity recognition, 3D point cloud and video frame labeling jobs. Use your labeling job task type to select one of the following ARNs and use it with this parameter when you create a labeling job. Replace aws-region with the Amazon Web Services Region you are creating your labeling job in. For example, replace aws-region with us-west-1 if you create a labeling job in US West (N. California).

Named Entity Recognition

Use the following HumanTaskUiArn for named entity recognition labeling jobs:

arn:aws:sagemaker:aws-region:394669845002:human-task-ui/NamedEntityRecognition

3D Point Cloud HumanTaskUiArns

Use this HumanTaskUiArn for 3D point cloud object detection and 3D point cloud object detection adjustment labeling jobs.

arn:aws:sagemaker:aws-region:394669845002:human-task-ui/PointCloudObjectDetection

Use this HumanTaskUiArn for 3D point cloud object tracking and 3D point cloud object tracking adjustment labeling jobs.

arn:aws:sagemaker:aws-region:394669845002:human-task-ui/PointCloudObjectTracking

Use this HumanTaskUiArn for 3D point cloud semantic segmentation and 3D point cloud semantic segmentation adjustment labeling jobs.

arn:aws:sagemaker:aws-region:394669845002:human-task-ui/PointCloudSemanticSegmentation

Video Frame HumanTaskUiArns

Use this HumanTaskUiArn for video frame object detection and video frame object detection adjustment labeling jobs.

arn:aws:sagemaker:region:394669845002:human-task-ui/VideoObjectDetection

Use this HumanTaskUiArn for video frame object tracking and video frame object tracking adjustment labeling jobs.

arn:aws:sagemaker:aws-region:394669845002:human-task-ui/VideoObjectTracking

uiConfig_uiTemplateS3Uri :: Lens' UiConfig (Maybe Text)

The Amazon S3 bucket location of the UI template, or worker task template. This is the template used to render the worker UI and tools for labeling job tasks. For more information about the contents of a UI template, see Creating Your Custom Labeling Task Template.

uiConfig_humanTaskUiArn :: Lens' UiConfig (Maybe Text)

The ARN of the worker task template used to render the worker UI and tools for labeling job tasks.

Use this parameter when you are creating a labeling job for named entity recognition, 3D point cloud and video frame labeling jobs. Use your labeling job task type to select one of the following ARNs and use it with this parameter when you create a labeling job. Replace aws-region with the Amazon Web Services Region you are creating your labeling job in. For example, replace aws-region with us-west-1 if you create a labeling job in US West (N. California).

Named Entity Recognition

Use the following HumanTaskUiArn for named entity recognition labeling jobs:

arn:aws:sagemaker:aws-region:394669845002:human-task-ui/NamedEntityRecognition

3D Point Cloud HumanTaskUiArns

Use this HumanTaskUiArn for 3D point cloud object detection and 3D point cloud object detection adjustment labeling jobs.

arn:aws:sagemaker:aws-region:394669845002:human-task-ui/PointCloudObjectDetection

Use this HumanTaskUiArn for 3D point cloud object tracking and 3D point cloud object tracking adjustment labeling jobs.

arn:aws:sagemaker:aws-region:394669845002:human-task-ui/PointCloudObjectTracking

Use this HumanTaskUiArn for 3D point cloud semantic segmentation and 3D point cloud semantic segmentation adjustment labeling jobs.

arn:aws:sagemaker:aws-region:394669845002:human-task-ui/PointCloudSemanticSegmentation

Video Frame HumanTaskUiArns

Use this HumanTaskUiArn for video frame object detection and video frame object detection adjustment labeling jobs.

arn:aws:sagemaker:region:394669845002:human-task-ui/VideoObjectDetection

Use this HumanTaskUiArn for video frame object tracking and video frame object tracking adjustment labeling jobs.

arn:aws:sagemaker:aws-region:394669845002:human-task-ui/VideoObjectTracking