Network.AWS.SageMaker.Types.UserContext

Description

Documentation

data UserContext

Information about the user who created or modified an experiment, trial, trial component, or project.

See: newUserContext smart constructor.

Constructors

UserContext'

Fields

userProfileName :: Maybe Text

The name of the user's profile.

domainId :: Maybe Text

The domain associated with the user.

userProfileArn :: Maybe Text

The Amazon Resource Name (ARN) of the user's profile.

newUserContext :: UserContext

Create a value of UserContext with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userProfileName:UserContext', userContext_userProfileName - The name of the user's profile.

$sel:domainId:UserContext', userContext_domainId - The domain associated with the user.

$sel:userProfileArn:UserContext', userContext_userProfileArn - The Amazon Resource Name (ARN) of the user's profile.

userContext_userProfileName :: Lens' UserContext (Maybe Text)

The name of the user's profile.

userContext_domainId :: Lens' UserContext (Maybe Text)

The domain associated with the user.

userContext_userProfileArn :: Lens' UserContext (Maybe Text)

The Amazon Resource Name (ARN) of the user's profile.