Network.AWS.SageMaker.Types.UserSettings

Description

Documentation

data UserSettings

A collection of settings that apply to users of Amazon SageMaker Studio. These settings are specified when the CreateUserProfile API is called, and as DefaultUserSettings when the CreateDomain API is called.

SecurityGroups is aggregated when specified in both calls. For all other settings in UserSettings, the values specified in CreateUserProfile take precedence over those specified in CreateDomain.

See: newUserSettings smart constructor.

Constructors

UserSettings'

Fields

kernelGatewayAppSettings :: Maybe KernelGatewayAppSettings

The kernel gateway app settings.

tensorBoardAppSettings :: Maybe TensorBoardAppSettings

The TensorBoard app settings.

securityGroups :: Maybe [Text]

The security groups for the Amazon Virtual Private Cloud (VPC) that Studio uses for communication.

Optional when the CreateDomain.AppNetworkAccessType parameter is set to PublicInternetOnly.

Required when the CreateDomain.AppNetworkAccessType parameter is set to VpcOnly.

Amazon SageMaker adds a security group to allow NFS traffic from SageMaker Studio. Therefore, the number of security groups that you can specify is one less than the maximum number shown.

jupyterServerAppSettings :: Maybe JupyterServerAppSettings

The Jupyter server's app settings.

executionRole :: Maybe Text

The execution role for the user.

sharingSettings :: Maybe SharingSettings

Specifies options for sharing SageMaker Studio notebooks.

newUserSettings :: UserSettings

Create a value of UserSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kernelGatewayAppSettings:UserSettings', userSettings_kernelGatewayAppSettings - The kernel gateway app settings.

$sel:tensorBoardAppSettings:UserSettings', userSettings_tensorBoardAppSettings - The TensorBoard app settings.

$sel:securityGroups:UserSettings', userSettings_securityGroups - The security groups for the Amazon Virtual Private Cloud (VPC) that Studio uses for communication.

Optional when the CreateDomain.AppNetworkAccessType parameter is set to PublicInternetOnly.

Required when the CreateDomain.AppNetworkAccessType parameter is set to VpcOnly.

Amazon SageMaker adds a security group to allow NFS traffic from SageMaker Studio. Therefore, the number of security groups that you can specify is one less than the maximum number shown.

$sel:jupyterServerAppSettings:UserSettings', userSettings_jupyterServerAppSettings - The Jupyter server's app settings.

$sel:executionRole:UserSettings', userSettings_executionRole - The execution role for the user.

$sel:sharingSettings:UserSettings', userSettings_sharingSettings - Specifies options for sharing SageMaker Studio notebooks.

userSettings_kernelGatewayAppSettings :: Lens' UserSettings (Maybe KernelGatewayAppSettings)

The kernel gateway app settings.

userSettings_tensorBoardAppSettings :: Lens' UserSettings (Maybe TensorBoardAppSettings)

The TensorBoard app settings.

userSettings_securityGroups :: Lens' UserSettings (Maybe [Text])

The security groups for the Amazon Virtual Private Cloud (VPC) that Studio uses for communication.

Optional when the CreateDomain.AppNetworkAccessType parameter is set to PublicInternetOnly.

Required when the CreateDomain.AppNetworkAccessType parameter is set to VpcOnly.

Amazon SageMaker adds a security group to allow NFS traffic from SageMaker Studio. Therefore, the number of security groups that you can specify is one less than the maximum number shown.

userSettings_jupyterServerAppSettings :: Lens' UserSettings (Maybe JupyterServerAppSettings)

The Jupyter server's app settings.

userSettings_executionRole :: Lens' UserSettings (Maybe Text)

The execution role for the user.

userSettings_sharingSettings :: Lens' UserSettings (Maybe SharingSettings)

Specifies options for sharing SageMaker Studio notebooks.