Network.AWS.SageMaker.Types.Workforce

Description

Documentation

data Workforce

A single private workforce, which is automatically created when you create your first private work team. You can create one private work force in each Amazon Web Services Region. By default, any workforce-related API operation used in a specific region will apply to the workforce created in that region. To learn how to create a private workforce, see Create a Private Workforce.

See: newWorkforce smart constructor.

Constructors

Workforce'

Fields

lastUpdatedDate :: Maybe POSIX

The most recent date that was used to successfully add one or more IP address ranges (CIDRs) to a private workforce's allow list.

createDate :: Maybe POSIX

The date that the workforce is created.

subDomain :: Maybe Text

The subdomain for your OIDC Identity Provider.

sourceIpConfig :: Maybe SourceIpConfig

A list of one to ten IP address ranges (CIDRs) to be added to the workforce allow list. By default, a workforce isn't restricted to specific IP addresses.

oidcConfig :: Maybe OidcConfigForResponse

The configuration of an OIDC Identity Provider (IdP) private workforce.

cognitoConfig :: Maybe CognitoConfig

The configuration of an Amazon Cognito workforce. A single Cognito workforce is created using and corresponds to a single Amazon Cognito user pool.

workforceName :: Text

The name of the private workforce.

workforceArn :: Text

The Amazon Resource Name (ARN) of the private workforce.

newWorkforce

Arguments

:: Text
$sel:workforceName:Workforce'

-> Text
$sel:workforceArn:Workforce'

-> Workforce

Create a value of Workforce with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedDate:Workforce', workforce_lastUpdatedDate - The most recent date that was used to successfully add one or more IP address ranges (CIDRs) to a private workforce's allow list.

$sel:createDate:Workforce', workforce_createDate - The date that the workforce is created.

$sel:subDomain:Workforce', workforce_subDomain - The subdomain for your OIDC Identity Provider.

$sel:sourceIpConfig:Workforce', workforce_sourceIpConfig - A list of one to ten IP address ranges (CIDRs) to be added to the workforce allow list. By default, a workforce isn't restricted to specific IP addresses.

$sel:oidcConfig:Workforce', workforce_oidcConfig - The configuration of an OIDC Identity Provider (IdP) private workforce.

$sel:cognitoConfig:Workforce', workforce_cognitoConfig - The configuration of an Amazon Cognito workforce. A single Cognito workforce is created using and corresponds to a single Amazon Cognito user pool.

$sel:workforceName:Workforce', workforce_workforceName - The name of the private workforce.

$sel:workforceArn:Workforce', workforce_workforceArn - The Amazon Resource Name (ARN) of the private workforce.

workforce_lastUpdatedDate :: Lens' Workforce (Maybe UTCTime)

The most recent date that was used to successfully add one or more IP address ranges (CIDRs) to a private workforce's allow list.

workforce_createDate :: Lens' Workforce (Maybe UTCTime)

The date that the workforce is created.

workforce_subDomain :: Lens' Workforce (Maybe Text)

The subdomain for your OIDC Identity Provider.

workforce_sourceIpConfig :: Lens' Workforce (Maybe SourceIpConfig)

A list of one to ten IP address ranges (CIDRs) to be added to the workforce allow list. By default, a workforce isn't restricted to specific IP addresses.

workforce_oidcConfig :: Lens' Workforce (Maybe OidcConfigForResponse)

The configuration of an OIDC Identity Provider (IdP) private workforce.

workforce_cognitoConfig :: Lens' Workforce (Maybe CognitoConfig)

The configuration of an Amazon Cognito workforce. A single Cognito workforce is created using and corresponds to a single Amazon Cognito user pool.

workforce_workforceName :: Lens' Workforce Text

The name of the private workforce.

workforce_workforceArn :: Lens' Workforce Text

The Amazon Resource Name (ARN) of the private workforce.