Network.AWS.SageMaker.UpdateAction

Description

Updates an action.

Creating a Request

data UpdateAction

See: newUpdateAction smart constructor.

Constructors

UpdateAction'

Fields

propertiesToRemove :: Maybe [Text]

A list of properties to remove.

status :: Maybe ActionStatus

The new status for the action.

properties :: Maybe (HashMap Text Text)

The new list of properties. Overwrites the current property list.

description :: Maybe Text

The new description for the action.

actionName :: Text

The name of the action to update.

newUpdateAction

Arguments

:: Text
$sel:actionName:UpdateAction'

-> UpdateAction

Create a value of UpdateAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:propertiesToRemove:UpdateAction', updateAction_propertiesToRemove - A list of properties to remove.

$sel:status:UpdateAction', updateAction_status - The new status for the action.

$sel:properties:UpdateAction', updateAction_properties - The new list of properties. Overwrites the current property list.

$sel:description:UpdateAction', updateAction_description - The new description for the action.

$sel:actionName:UpdateAction', updateAction_actionName - The name of the action to update.

Request Lenses

updateAction_propertiesToRemove :: Lens' UpdateAction (Maybe [Text])

A list of properties to remove.

updateAction_status :: Lens' UpdateAction (Maybe ActionStatus)

The new status for the action.

updateAction_properties :: Lens' UpdateAction (Maybe (HashMap Text Text))

The new list of properties. Overwrites the current property list.

updateAction_description :: Lens' UpdateAction (Maybe Text)

The new description for the action.

updateAction_actionName :: Lens' UpdateAction Text

The name of the action to update.

Destructuring the Response

data UpdateActionResponse

See: newUpdateActionResponse smart constructor.

Constructors

UpdateActionResponse'

Fields

actionArn :: Maybe Text

The Amazon Resource Name (ARN) of the action.

httpStatus :: Int

The response's http status code.

newUpdateActionResponse

Arguments

:: Int
$sel:httpStatus:UpdateActionResponse'

-> UpdateActionResponse

Create a value of UpdateActionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actionArn:UpdateActionResponse', updateActionResponse_actionArn - The Amazon Resource Name (ARN) of the action.

$sel:httpStatus:UpdateActionResponse', updateActionResponse_httpStatus - The response's http status code.

Response Lenses

updateActionResponse_actionArn :: Lens' UpdateActionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the action.

updateActionResponse_httpStatus :: Lens' UpdateActionResponse Int

The response's http status code.