Network.AWS.SageMaker.UpdateImage

Description

Updates the properties of a SageMaker image. To change the image's tags, use the AddTags and DeleteTags APIs.

Creating a Request

data UpdateImage

See: newUpdateImage smart constructor.

Constructors

UpdateImage'

Fields

deleteProperties :: Maybe [Text]

A list of properties to delete. Only the Description and DisplayName properties can be deleted.

displayName :: Maybe Text

The new display name for the image.

description :: Maybe Text

The new description for the image.

roleArn :: Maybe Text

The new Amazon Resource Name (ARN) for the IAM role that enables Amazon SageMaker to perform tasks on your behalf.

imageName :: Text

The name of the image to update.

newUpdateImage

Arguments

:: Text
$sel:imageName:UpdateImage'

-> UpdateImage

Create a value of UpdateImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deleteProperties:UpdateImage', updateImage_deleteProperties - A list of properties to delete. Only the Description and DisplayName properties can be deleted.

$sel:displayName:UpdateImage', updateImage_displayName - The new display name for the image.

$sel:description:UpdateImage', updateImage_description - The new description for the image.

$sel:roleArn:UpdateImage', updateImage_roleArn - The new Amazon Resource Name (ARN) for the IAM role that enables Amazon SageMaker to perform tasks on your behalf.

$sel:imageName:UpdateImage', updateImage_imageName - The name of the image to update.

Request Lenses

updateImage_deleteProperties :: Lens' UpdateImage (Maybe [Text])

A list of properties to delete. Only the Description and DisplayName properties can be deleted.

updateImage_displayName :: Lens' UpdateImage (Maybe Text)

The new display name for the image.

updateImage_description :: Lens' UpdateImage (Maybe Text)

The new description for the image.

updateImage_roleArn :: Lens' UpdateImage (Maybe Text)

The new Amazon Resource Name (ARN) for the IAM role that enables Amazon SageMaker to perform tasks on your behalf.

updateImage_imageName :: Lens' UpdateImage Text

The name of the image to update.

Destructuring the Response

data UpdateImageResponse

See: newUpdateImageResponse smart constructor.

Constructors

UpdateImageResponse'

Fields

imageArn :: Maybe Text

The Amazon Resource Name (ARN) of the image.

httpStatus :: Int

The response's http status code.

newUpdateImageResponse

Arguments

:: Int
$sel:httpStatus:UpdateImageResponse'

-> UpdateImageResponse

Create a value of UpdateImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageArn:UpdateImageResponse', updateImageResponse_imageArn - The Amazon Resource Name (ARN) of the image.

$sel:httpStatus:UpdateImageResponse', updateImageResponse_httpStatus - The response's http status code.

Response Lenses

updateImageResponse_imageArn :: Lens' UpdateImageResponse (Maybe Text)

The Amazon Resource Name (ARN) of the image.

updateImageResponse_httpStatus :: Lens' UpdateImageResponse Int

The response's http status code.