Network.AWS.SageMaker.UpdateNotebookInstance

Description

Updates a notebook instance. NotebookInstance updates include upgrading or downgrading the ML compute instance used for your notebook instance to accommodate changes in your workload requirements.

Creating a Request

data UpdateNotebookInstance

See: newUpdateNotebookInstance smart constructor.

Constructors

UpdateNotebookInstance'

Fields

acceleratorTypes :: Maybe [NotebookInstanceAcceleratorType]

A list of the Elastic Inference (EI) instance types to associate with this notebook instance. Currently only one EI instance type can be associated with a notebook instance. For more information, see Using Elastic Inference in Amazon SageMaker.

disassociateAdditionalCodeRepositories :: Maybe Bool

A list of names or URLs of the default Git repositories to remove from this notebook instance. This operation is idempotent. If you specify a Git repository that is not associated with the notebook instance when you call this method, it does not throw an error.

additionalCodeRepositories :: Maybe [Text]

An array of up to three Git repositories to associate with the notebook instance. These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in Amazon Web Services CodeCommit or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

lifecycleConfigName :: Maybe Text

The name of a lifecycle configuration to associate with the notebook instance. For information about lifestyle configurations, see Step 2.1: (Optional) Customize a Notebook Instance.

disassociateLifecycleConfig :: Maybe Bool

Set to true to remove the notebook instance lifecycle configuration currently associated with the notebook instance. This operation is idempotent. If you specify a lifecycle configuration that is not associated with the notebook instance when you call this method, it does not throw an error.

disassociateDefaultCodeRepository :: Maybe Bool

The name or URL of the default Git repository to remove from this notebook instance. This operation is idempotent. If you specify a Git repository that is not associated with the notebook instance when you call this method, it does not throw an error.

instanceType :: Maybe InstanceType

The Amazon ML compute instance type.

defaultCodeRepository :: Maybe Text

The Git repository to associate with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in Amazon Web Services CodeCommit or in any other Git repository. When you open a notebook instance, it opens in the directory that contains this repository. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

volumeSizeInGB :: Maybe Natural

The size, in GB, of the ML storage volume to attach to the notebook instance. The default value is 5 GB. ML storage volumes are encrypted, so Amazon SageMaker can't determine the amount of available free space on the volume. Because of this, you can increase the volume size when you update a notebook instance, but you can't decrease the volume size. If you want to decrease the size of the ML storage volume in use, create a new notebook instance with the desired size.

rootAccess :: Maybe RootAccess

Whether root access is enabled or disabled for users of the notebook instance. The default value is Enabled.

If you set this to Disabled, users don't have root access on the notebook instance, but lifecycle configuration scripts still run with root permissions.

disassociateAcceleratorTypes :: Maybe Bool

A list of the Elastic Inference (EI) instance types to remove from this notebook instance. This operation is idempotent. If you specify an accelerator type that is not associated with the notebook instance when you call this method, it does not throw an error.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role that Amazon SageMaker can assume to access the notebook instance. For more information, see Amazon SageMaker Roles.

To be able to pass this role to Amazon SageMaker, the caller of this API must have the iam:PassRole permission.

notebookInstanceName :: Text

The name of the notebook instance to update.

newUpdateNotebookInstance

Arguments

:: Text
$sel:notebookInstanceName:UpdateNotebookInstance'

-> UpdateNotebookInstance

Create a value of UpdateNotebookInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceleratorTypes:UpdateNotebookInstance', updateNotebookInstance_acceleratorTypes - A list of the Elastic Inference (EI) instance types to associate with this notebook instance. Currently only one EI instance type can be associated with a notebook instance. For more information, see Using Elastic Inference in Amazon SageMaker.

$sel:disassociateAdditionalCodeRepositories:UpdateNotebookInstance', updateNotebookInstance_disassociateAdditionalCodeRepositories - A list of names or URLs of the default Git repositories to remove from this notebook instance. This operation is idempotent. If you specify a Git repository that is not associated with the notebook instance when you call this method, it does not throw an error.

$sel:additionalCodeRepositories:UpdateNotebookInstance', updateNotebookInstance_additionalCodeRepositories - An array of up to three Git repositories to associate with the notebook instance. These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in Amazon Web Services CodeCommit or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

$sel:lifecycleConfigName:UpdateNotebookInstance', updateNotebookInstance_lifecycleConfigName - The name of a lifecycle configuration to associate with the notebook instance. For information about lifestyle configurations, see Step 2.1: (Optional) Customize a Notebook Instance.

$sel:disassociateLifecycleConfig:UpdateNotebookInstance', updateNotebookInstance_disassociateLifecycleConfig - Set to true to remove the notebook instance lifecycle configuration currently associated with the notebook instance. This operation is idempotent. If you specify a lifecycle configuration that is not associated with the notebook instance when you call this method, it does not throw an error.

$sel:disassociateDefaultCodeRepository:UpdateNotebookInstance', updateNotebookInstance_disassociateDefaultCodeRepository - The name or URL of the default Git repository to remove from this notebook instance. This operation is idempotent. If you specify a Git repository that is not associated with the notebook instance when you call this method, it does not throw an error.

$sel:instanceType:UpdateNotebookInstance', updateNotebookInstance_instanceType - The Amazon ML compute instance type.

$sel:defaultCodeRepository:UpdateNotebookInstance', updateNotebookInstance_defaultCodeRepository - The Git repository to associate with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in Amazon Web Services CodeCommit or in any other Git repository. When you open a notebook instance, it opens in the directory that contains this repository. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

$sel:volumeSizeInGB:UpdateNotebookInstance', updateNotebookInstance_volumeSizeInGB - The size, in GB, of the ML storage volume to attach to the notebook instance. The default value is 5 GB. ML storage volumes are encrypted, so Amazon SageMaker can't determine the amount of available free space on the volume. Because of this, you can increase the volume size when you update a notebook instance, but you can't decrease the volume size. If you want to decrease the size of the ML storage volume in use, create a new notebook instance with the desired size.

$sel:rootAccess:UpdateNotebookInstance', updateNotebookInstance_rootAccess - Whether root access is enabled or disabled for users of the notebook instance. The default value is Enabled.

If you set this to Disabled, users don't have root access on the notebook instance, but lifecycle configuration scripts still run with root permissions.

$sel:disassociateAcceleratorTypes:UpdateNotebookInstance', updateNotebookInstance_disassociateAcceleratorTypes - A list of the Elastic Inference (EI) instance types to remove from this notebook instance. This operation is idempotent. If you specify an accelerator type that is not associated with the notebook instance when you call this method, it does not throw an error.

$sel:roleArn:UpdateNotebookInstance', updateNotebookInstance_roleArn - The Amazon Resource Name (ARN) of the IAM role that Amazon SageMaker can assume to access the notebook instance. For more information, see Amazon SageMaker Roles.

To be able to pass this role to Amazon SageMaker, the caller of this API must have the iam:PassRole permission.

$sel:notebookInstanceName:UpdateNotebookInstance', updateNotebookInstance_notebookInstanceName - The name of the notebook instance to update.

Request Lenses

updateNotebookInstance_acceleratorTypes :: Lens' UpdateNotebookInstance (Maybe [NotebookInstanceAcceleratorType])

A list of the Elastic Inference (EI) instance types to associate with this notebook instance. Currently only one EI instance type can be associated with a notebook instance. For more information, see Using Elastic Inference in Amazon SageMaker.

updateNotebookInstance_disassociateAdditionalCodeRepositories :: Lens' UpdateNotebookInstance (Maybe Bool)

A list of names or URLs of the default Git repositories to remove from this notebook instance. This operation is idempotent. If you specify a Git repository that is not associated with the notebook instance when you call this method, it does not throw an error.

updateNotebookInstance_additionalCodeRepositories :: Lens' UpdateNotebookInstance (Maybe [Text])

An array of up to three Git repositories to associate with the notebook instance. These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in Amazon Web Services CodeCommit or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

updateNotebookInstance_lifecycleConfigName :: Lens' UpdateNotebookInstance (Maybe Text)

The name of a lifecycle configuration to associate with the notebook instance. For information about lifestyle configurations, see Step 2.1: (Optional) Customize a Notebook Instance.

updateNotebookInstance_disassociateLifecycleConfig :: Lens' UpdateNotebookInstance (Maybe Bool)

Set to true to remove the notebook instance lifecycle configuration currently associated with the notebook instance. This operation is idempotent. If you specify a lifecycle configuration that is not associated with the notebook instance when you call this method, it does not throw an error.

updateNotebookInstance_disassociateDefaultCodeRepository :: Lens' UpdateNotebookInstance (Maybe Bool)

The name or URL of the default Git repository to remove from this notebook instance. This operation is idempotent. If you specify a Git repository that is not associated with the notebook instance when you call this method, it does not throw an error.

updateNotebookInstance_instanceType :: Lens' UpdateNotebookInstance (Maybe InstanceType)

The Amazon ML compute instance type.

updateNotebookInstance_defaultCodeRepository :: Lens' UpdateNotebookInstance (Maybe Text)

The Git repository to associate with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in Amazon Web Services CodeCommit or in any other Git repository. When you open a notebook instance, it opens in the directory that contains this repository. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

updateNotebookInstance_volumeSizeInGB :: Lens' UpdateNotebookInstance (Maybe Natural)

The size, in GB, of the ML storage volume to attach to the notebook instance. The default value is 5 GB. ML storage volumes are encrypted, so Amazon SageMaker can't determine the amount of available free space on the volume. Because of this, you can increase the volume size when you update a notebook instance, but you can't decrease the volume size. If you want to decrease the size of the ML storage volume in use, create a new notebook instance with the desired size.

updateNotebookInstance_rootAccess :: Lens' UpdateNotebookInstance (Maybe RootAccess)

Whether root access is enabled or disabled for users of the notebook instance. The default value is Enabled.

If you set this to Disabled, users don't have root access on the notebook instance, but lifecycle configuration scripts still run with root permissions.

updateNotebookInstance_disassociateAcceleratorTypes :: Lens' UpdateNotebookInstance (Maybe Bool)

A list of the Elastic Inference (EI) instance types to remove from this notebook instance. This operation is idempotent. If you specify an accelerator type that is not associated with the notebook instance when you call this method, it does not throw an error.

updateNotebookInstance_roleArn :: Lens' UpdateNotebookInstance (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that Amazon SageMaker can assume to access the notebook instance. For more information, see Amazon SageMaker Roles.

To be able to pass this role to Amazon SageMaker, the caller of this API must have the iam:PassRole permission.

updateNotebookInstance_notebookInstanceName :: Lens' UpdateNotebookInstance Text

The name of the notebook instance to update.

Destructuring the Response

data UpdateNotebookInstanceResponse

See: newUpdateNotebookInstanceResponse smart constructor.

Constructors

UpdateNotebookInstanceResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateNotebookInstanceResponse

Arguments

:: Int
$sel:httpStatus:UpdateNotebookInstanceResponse'

-> UpdateNotebookInstanceResponse

Create a value of UpdateNotebookInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateNotebookInstanceResponse', updateNotebookInstanceResponse_httpStatus - The response's http status code.

Response Lenses

updateNotebookInstanceResponse_httpStatus :: Lens' UpdateNotebookInstanceResponse Int

The response's http status code.