Network.AWS.SageMaker.UpdateTrainingJob

Description

Update a model training job to request a new Debugger profiling configuration.

Creating a Request

data UpdateTrainingJob

See: newUpdateTrainingJob smart constructor.

Constructors

UpdateTrainingJob'

Fields

profilerConfig :: Maybe ProfilerConfigForUpdate

Configuration information for Debugger system monitoring, framework profiling, and storage paths.

profilerRuleConfigurations :: Maybe [ProfilerRuleConfiguration]

Configuration information for Debugger rules for profiling system and framework metrics.

trainingJobName :: Text

The name of a training job to update the Debugger profiling configuration.

newUpdateTrainingJob

Arguments

:: Text
$sel:trainingJobName:UpdateTrainingJob'

-> UpdateTrainingJob

Create a value of UpdateTrainingJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:profilerConfig:UpdateTrainingJob', updateTrainingJob_profilerConfig - Configuration information for Debugger system monitoring, framework profiling, and storage paths.

$sel:profilerRuleConfigurations:UpdateTrainingJob', updateTrainingJob_profilerRuleConfigurations - Configuration information for Debugger rules for profiling system and framework metrics.

$sel:trainingJobName:UpdateTrainingJob', updateTrainingJob_trainingJobName - The name of a training job to update the Debugger profiling configuration.

Request Lenses

updateTrainingJob_profilerConfig :: Lens' UpdateTrainingJob (Maybe ProfilerConfigForUpdate)

Configuration information for Debugger system monitoring, framework profiling, and storage paths.

updateTrainingJob_profilerRuleConfigurations :: Lens' UpdateTrainingJob (Maybe [ProfilerRuleConfiguration])

Configuration information for Debugger rules for profiling system and framework metrics.

updateTrainingJob_trainingJobName :: Lens' UpdateTrainingJob Text

The name of a training job to update the Debugger profiling configuration.

Destructuring the Response

data UpdateTrainingJobResponse

See: newUpdateTrainingJobResponse smart constructor.

Constructors

UpdateTrainingJobResponse'

Fields

httpStatus :: Int

The response's http status code.

trainingJobArn :: Text

The Amazon Resource Name (ARN) of the training job.

newUpdateTrainingJobResponse

Arguments

:: Int
$sel:httpStatus:UpdateTrainingJobResponse'

-> Text
$sel:trainingJobArn:UpdateTrainingJobResponse'

-> UpdateTrainingJobResponse

Create a value of UpdateTrainingJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateTrainingJobResponse', updateTrainingJobResponse_httpStatus - The response's http status code.

$sel:trainingJobArn:UpdateTrainingJobResponse', updateTrainingJobResponse_trainingJobArn - The Amazon Resource Name (ARN) of the training job.

Response Lenses

updateTrainingJobResponse_httpStatus :: Lens' UpdateTrainingJobResponse Int

The response's http status code.

updateTrainingJobResponse_trainingJobArn :: Lens' UpdateTrainingJobResponse Text

The Amazon Resource Name (ARN) of the training job.