Network.AWS.SageMaker.UpdateTrialComponent

Description

Updates one or more properties of a trial component.

Creating a Request

data UpdateTrialComponent

See: newUpdateTrialComponent smart constructor.

Constructors

UpdateTrialComponent'

Fields

parametersToRemove :: Maybe [Text]

The hyperparameters to remove from the component.

outputArtifactsToRemove :: Maybe [Text]

The output artifacts to remove from the component.

status :: Maybe TrialComponentStatus

The new status of the component.

inputArtifactsToRemove :: Maybe [Text]

The input artifacts to remove from the component.

startTime :: Maybe POSIX

When the component started.

endTime :: Maybe POSIX

When the component ended.

inputArtifacts :: Maybe (HashMap Text TrialComponentArtifact)

Replaces all of the component's input artifacts with the specified artifacts.

displayName :: Maybe Text

The name of the component as displayed. The name doesn't need to be unique. If DisplayName isn't specified, TrialComponentName is displayed.

outputArtifacts :: Maybe (HashMap Text TrialComponentArtifact)

Replaces all of the component's output artifacts with the specified artifacts.

parameters :: Maybe (HashMap Text TrialComponentParameterValue)

Replaces all of the component's hyperparameters with the specified hyperparameters.

trialComponentName :: Text

The name of the component to update.

newUpdateTrialComponent

Arguments

:: Text
$sel:trialComponentName:UpdateTrialComponent'

-> UpdateTrialComponent

Create a value of UpdateTrialComponent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parametersToRemove:UpdateTrialComponent', updateTrialComponent_parametersToRemove - The hyperparameters to remove from the component.

$sel:outputArtifactsToRemove:UpdateTrialComponent', updateTrialComponent_outputArtifactsToRemove - The output artifacts to remove from the component.

$sel:status:UpdateTrialComponent', updateTrialComponent_status - The new status of the component.

$sel:inputArtifactsToRemove:UpdateTrialComponent', updateTrialComponent_inputArtifactsToRemove - The input artifacts to remove from the component.

$sel:startTime:UpdateTrialComponent', updateTrialComponent_startTime - When the component started.

$sel:endTime:UpdateTrialComponent', updateTrialComponent_endTime - When the component ended.

$sel:inputArtifacts:UpdateTrialComponent', updateTrialComponent_inputArtifacts - Replaces all of the component's input artifacts with the specified artifacts.

$sel:displayName:UpdateTrialComponent', updateTrialComponent_displayName - The name of the component as displayed. The name doesn't need to be unique. If DisplayName isn't specified, TrialComponentName is displayed.

$sel:outputArtifacts:UpdateTrialComponent', updateTrialComponent_outputArtifacts - Replaces all of the component's output artifacts with the specified artifacts.

$sel:parameters:UpdateTrialComponent', updateTrialComponent_parameters - Replaces all of the component's hyperparameters with the specified hyperparameters.

$sel:trialComponentName:UpdateTrialComponent', updateTrialComponent_trialComponentName - The name of the component to update.

Request Lenses

updateTrialComponent_parametersToRemove :: Lens' UpdateTrialComponent (Maybe [Text])

The hyperparameters to remove from the component.

updateTrialComponent_outputArtifactsToRemove :: Lens' UpdateTrialComponent (Maybe [Text])

The output artifacts to remove from the component.

updateTrialComponent_status :: Lens' UpdateTrialComponent (Maybe TrialComponentStatus)

The new status of the component.

updateTrialComponent_inputArtifactsToRemove :: Lens' UpdateTrialComponent (Maybe [Text])

The input artifacts to remove from the component.

updateTrialComponent_startTime :: Lens' UpdateTrialComponent (Maybe UTCTime)

When the component started.

updateTrialComponent_endTime :: Lens' UpdateTrialComponent (Maybe UTCTime)

When the component ended.

updateTrialComponent_inputArtifacts :: Lens' UpdateTrialComponent (Maybe (HashMap Text TrialComponentArtifact))

Replaces all of the component's input artifacts with the specified artifacts.

updateTrialComponent_displayName :: Lens' UpdateTrialComponent (Maybe Text)

The name of the component as displayed. The name doesn't need to be unique. If DisplayName isn't specified, TrialComponentName is displayed.

updateTrialComponent_outputArtifacts :: Lens' UpdateTrialComponent (Maybe (HashMap Text TrialComponentArtifact))

Replaces all of the component's output artifacts with the specified artifacts.

updateTrialComponent_parameters :: Lens' UpdateTrialComponent (Maybe (HashMap Text TrialComponentParameterValue))

Replaces all of the component's hyperparameters with the specified hyperparameters.

updateTrialComponent_trialComponentName :: Lens' UpdateTrialComponent Text

The name of the component to update.

Destructuring the Response

data UpdateTrialComponentResponse

See: newUpdateTrialComponentResponse smart constructor.

Constructors

UpdateTrialComponentResponse'

Fields

trialComponentArn :: Maybe Text

The Amazon Resource Name (ARN) of the trial component.

httpStatus :: Int

The response's http status code.

newUpdateTrialComponentResponse

Arguments

:: Int
$sel:httpStatus:UpdateTrialComponentResponse'

-> UpdateTrialComponentResponse

Create a value of UpdateTrialComponentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trialComponentArn:UpdateTrialComponentResponse', updateTrialComponentResponse_trialComponentArn - The Amazon Resource Name (ARN) of the trial component.

$sel:httpStatus:UpdateTrialComponentResponse', updateTrialComponentResponse_httpStatus - The response's http status code.

Response Lenses

updateTrialComponentResponse_trialComponentArn :: Lens' UpdateTrialComponentResponse (Maybe Text)

The Amazon Resource Name (ARN) of the trial component.

updateTrialComponentResponse_httpStatus :: Lens' UpdateTrialComponentResponse Int

The response's http status code.