Network.AWS.SageMakerA2IRuntime.DescribeHumanLoop

Description

Returns information about the specified human loop. If the human loop was deleted, this operation will return a ResourceNotFoundException error.

Creating a Request

data DescribeHumanLoop

See: newDescribeHumanLoop smart constructor.

Constructors

DescribeHumanLoop'

Fields

humanLoopName :: Text

The name of the human loop that you want information about.

newDescribeHumanLoop

Arguments

:: Text
$sel:humanLoopName:DescribeHumanLoop'

-> DescribeHumanLoop

Create a value of DescribeHumanLoop with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:humanLoopName:DescribeHumanLoop', describeHumanLoop_humanLoopName - The name of the human loop that you want information about.

Request Lenses

describeHumanLoop_humanLoopName :: Lens' DescribeHumanLoop Text

The name of the human loop that you want information about.

Destructuring the Response

data DescribeHumanLoopResponse

See: newDescribeHumanLoopResponse smart constructor.

Constructors

DescribeHumanLoopResponse'

Fields

failureReason :: Maybe Text

The reason why a human loop failed. The failure reason is returned when the status of the human loop is Failed.

failureCode :: Maybe Text

A failure code that identifies the type of failure.

Possible values: ValidationError, Expired, InternalError

humanLoopOutput :: Maybe HumanLoopOutput

An object that contains information about the output of the human loop.

httpStatus :: Int

The response's http status code.

creationTime :: POSIX

The creation time when Amazon Augmented AI created the human loop.

humanLoopStatus :: HumanLoopStatus

The status of the human loop.

humanLoopName :: Text

The name of the human loop. The name must be lowercase, unique within the Region in your account, and can have up to 63 characters. Valid characters: a-z, 0-9, and - (hyphen).

humanLoopArn :: Text

The Amazon Resource Name (ARN) of the human loop.

flowDefinitionArn :: Text

The Amazon Resource Name (ARN) of the flow definition.

newDescribeHumanLoopResponse

Arguments

:: Int
$sel:httpStatus:DescribeHumanLoopResponse'

-> UTCTime
$sel:creationTime:DescribeHumanLoopResponse'

-> HumanLoopStatus
$sel:humanLoopStatus:DescribeHumanLoopResponse'

-> Text
$sel:humanLoopName:DescribeHumanLoop'

-> Text
$sel:humanLoopArn:DescribeHumanLoopResponse'

-> Text
$sel:flowDefinitionArn:DescribeHumanLoopResponse'

-> DescribeHumanLoopResponse

Create a value of DescribeHumanLoopResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:DescribeHumanLoopResponse', describeHumanLoopResponse_failureReason - The reason why a human loop failed. The failure reason is returned when the status of the human loop is Failed.

$sel:failureCode:DescribeHumanLoopResponse', describeHumanLoopResponse_failureCode - A failure code that identifies the type of failure.

Possible values: ValidationError, Expired, InternalError

$sel:humanLoopOutput:DescribeHumanLoopResponse', describeHumanLoopResponse_humanLoopOutput - An object that contains information about the output of the human loop.

$sel:httpStatus:DescribeHumanLoopResponse', describeHumanLoopResponse_httpStatus - The response's http status code.

$sel:creationTime:DescribeHumanLoopResponse', describeHumanLoopResponse_creationTime - The creation time when Amazon Augmented AI created the human loop.

$sel:humanLoopStatus:DescribeHumanLoopResponse', describeHumanLoopResponse_humanLoopStatus - The status of the human loop.

$sel:humanLoopName:DescribeHumanLoop', describeHumanLoopResponse_humanLoopName - The name of the human loop. The name must be lowercase, unique within the Region in your account, and can have up to 63 characters. Valid characters: a-z, 0-9, and - (hyphen).

$sel:humanLoopArn:DescribeHumanLoopResponse', describeHumanLoopResponse_humanLoopArn - The Amazon Resource Name (ARN) of the human loop.

$sel:flowDefinitionArn:DescribeHumanLoopResponse', describeHumanLoopResponse_flowDefinitionArn - The Amazon Resource Name (ARN) of the flow definition.

Response Lenses

describeHumanLoopResponse_failureReason :: Lens' DescribeHumanLoopResponse (Maybe Text)

The reason why a human loop failed. The failure reason is returned when the status of the human loop is Failed.

describeHumanLoopResponse_failureCode :: Lens' DescribeHumanLoopResponse (Maybe Text)

A failure code that identifies the type of failure.

Possible values: ValidationError, Expired, InternalError

describeHumanLoopResponse_humanLoopOutput :: Lens' DescribeHumanLoopResponse (Maybe HumanLoopOutput)

An object that contains information about the output of the human loop.

describeHumanLoopResponse_httpStatus :: Lens' DescribeHumanLoopResponse Int

The response's http status code.

describeHumanLoopResponse_creationTime :: Lens' DescribeHumanLoopResponse UTCTime

The creation time when Amazon Augmented AI created the human loop.

describeHumanLoopResponse_humanLoopStatus :: Lens' DescribeHumanLoopResponse HumanLoopStatus

The status of the human loop.

describeHumanLoopResponse_humanLoopName :: Lens' DescribeHumanLoopResponse Text

The name of the human loop. The name must be lowercase, unique within the Region in your account, and can have up to 63 characters. Valid characters: a-z, 0-9, and - (hyphen).

describeHumanLoopResponse_humanLoopArn :: Lens' DescribeHumanLoopResponse Text

The Amazon Resource Name (ARN) of the human loop.

describeHumanLoopResponse_flowDefinitionArn :: Lens' DescribeHumanLoopResponse Text

The Amazon Resource Name (ARN) of the flow definition.