Network.AWS.SageMakerA2IRuntime.ListHumanLoops

Description

Returns information about human loops, given the specified parameters. If a human loop was deleted, it will not be included.

This operation returns paginated results.

Creating a Request

data ListHumanLoops

See: newListHumanLoops smart constructor.

Constructors

ListHumanLoops'

Fields

creationTimeAfter :: Maybe POSIX

(Optional) The timestamp of the date when you want the human loops to begin in ISO 8601 format. For example, 2020-02-24.

nextToken :: Maybe Text

A token to display the next page of results.

sortOrder :: Maybe SortOrder

Optional. The order for displaying results. Valid values: Ascending and Descending.

creationTimeBefore :: Maybe POSIX

(Optional) The timestamp of the date before which you want the human loops to begin in ISO 8601 format. For example, 2020-02-24.

maxResults :: Maybe Natural

The total number of items to return. If the total number of available items is more than the value specified in MaxResults, then a NextToken is returned in the output. You can use this token to display the next page of results.

flowDefinitionArn :: Text

The Amazon Resource Name (ARN) of a flow definition.

newListHumanLoops

Arguments

:: Text
$sel:flowDefinitionArn:ListHumanLoops'

-> ListHumanLoops

Create a value of ListHumanLoops with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTimeAfter:ListHumanLoops', listHumanLoops_creationTimeAfter - (Optional) The timestamp of the date when you want the human loops to begin in ISO 8601 format. For example, 2020-02-24.

$sel:nextToken:ListHumanLoops', listHumanLoops_nextToken - A token to display the next page of results.

$sel:sortOrder:ListHumanLoops', listHumanLoops_sortOrder - Optional. The order for displaying results. Valid values: Ascending and Descending.

$sel:creationTimeBefore:ListHumanLoops', listHumanLoops_creationTimeBefore - (Optional) The timestamp of the date before which you want the human loops to begin in ISO 8601 format. For example, 2020-02-24.

$sel:maxResults:ListHumanLoops', listHumanLoops_maxResults - The total number of items to return. If the total number of available items is more than the value specified in MaxResults, then a NextToken is returned in the output. You can use this token to display the next page of results.

$sel:flowDefinitionArn:ListHumanLoops', listHumanLoops_flowDefinitionArn - The Amazon Resource Name (ARN) of a flow definition.

Request Lenses

listHumanLoops_creationTimeAfter :: Lens' ListHumanLoops (Maybe UTCTime)

(Optional) The timestamp of the date when you want the human loops to begin in ISO 8601 format. For example, 2020-02-24.

listHumanLoops_nextToken :: Lens' ListHumanLoops (Maybe Text)

A token to display the next page of results.

listHumanLoops_sortOrder :: Lens' ListHumanLoops (Maybe SortOrder)

Optional. The order for displaying results. Valid values: Ascending and Descending.

listHumanLoops_creationTimeBefore :: Lens' ListHumanLoops (Maybe UTCTime)

(Optional) The timestamp of the date before which you want the human loops to begin in ISO 8601 format. For example, 2020-02-24.

listHumanLoops_maxResults :: Lens' ListHumanLoops (Maybe Natural)

The total number of items to return. If the total number of available items is more than the value specified in MaxResults, then a NextToken is returned in the output. You can use this token to display the next page of results.

listHumanLoops_flowDefinitionArn :: Lens' ListHumanLoops Text

The Amazon Resource Name (ARN) of a flow definition.

Destructuring the Response

data ListHumanLoopsResponse

See: newListHumanLoopsResponse smart constructor.

Constructors

ListHumanLoopsResponse'

Fields

nextToken :: Maybe Text

A token to display the next page of results.

httpStatus :: Int

The response's http status code.

humanLoopSummaries :: [HumanLoopSummary]

An array of objects that contain information about the human loops.

newListHumanLoopsResponse

Arguments

:: Int
$sel:httpStatus:ListHumanLoopsResponse'

-> ListHumanLoopsResponse

Create a value of ListHumanLoopsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListHumanLoops', listHumanLoopsResponse_nextToken - A token to display the next page of results.

$sel:httpStatus:ListHumanLoopsResponse', listHumanLoopsResponse_httpStatus - The response's http status code.

$sel:humanLoopSummaries:ListHumanLoopsResponse', listHumanLoopsResponse_humanLoopSummaries - An array of objects that contain information about the human loops.

Response Lenses

listHumanLoopsResponse_nextToken :: Lens' ListHumanLoopsResponse (Maybe Text)

A token to display the next page of results.

listHumanLoopsResponse_httpStatus :: Lens' ListHumanLoopsResponse Int

The response's http status code.

listHumanLoopsResponse_humanLoopSummaries :: Lens' ListHumanLoopsResponse [HumanLoopSummary]

An array of objects that contain information about the human loops.