Network.AWS.SageMakerA2IRuntime.StartHumanLoop

Description

Starts a human loop, provided that at least one activation condition is met.

Creating a Request

data StartHumanLoop

See: newStartHumanLoop smart constructor.

Constructors

StartHumanLoop'

Fields

dataAttributes :: Maybe HumanLoopDataAttributes

Attributes of the specified data. Use DataAttributes to specify if your data is free of personally identifiable information and/or free of adult content.

humanLoopName :: Text

The name of the human loop.

flowDefinitionArn :: Text

The Amazon Resource Name (ARN) of the flow definition associated with this human loop.

humanLoopInput :: HumanLoopInput

An object that contains information about the human loop.

newStartHumanLoop

Arguments

:: Text
$sel:humanLoopName:StartHumanLoop'

-> Text
$sel:flowDefinitionArn:StartHumanLoop'

-> HumanLoopInput
$sel:humanLoopInput:StartHumanLoop'

-> StartHumanLoop

Create a value of StartHumanLoop with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataAttributes:StartHumanLoop', startHumanLoop_dataAttributes - Attributes of the specified data. Use DataAttributes to specify if your data is free of personally identifiable information and/or free of adult content.

$sel:humanLoopName:StartHumanLoop', startHumanLoop_humanLoopName - The name of the human loop.

$sel:flowDefinitionArn:StartHumanLoop', startHumanLoop_flowDefinitionArn - The Amazon Resource Name (ARN) of the flow definition associated with this human loop.

$sel:humanLoopInput:StartHumanLoop', startHumanLoop_humanLoopInput - An object that contains information about the human loop.

Request Lenses

startHumanLoop_dataAttributes :: Lens' StartHumanLoop (Maybe HumanLoopDataAttributes)

Attributes of the specified data. Use DataAttributes to specify if your data is free of personally identifiable information and/or free of adult content.

startHumanLoop_humanLoopName :: Lens' StartHumanLoop Text

The name of the human loop.

startHumanLoop_flowDefinitionArn :: Lens' StartHumanLoop Text

The Amazon Resource Name (ARN) of the flow definition associated with this human loop.

startHumanLoop_humanLoopInput :: Lens' StartHumanLoop HumanLoopInput

An object that contains information about the human loop.

Destructuring the Response

data StartHumanLoopResponse

See: newStartHumanLoopResponse smart constructor.

Constructors

StartHumanLoopResponse'

Fields

humanLoopArn :: Maybe Text

The Amazon Resource Name (ARN) of the human loop.

httpStatus :: Int

The response's http status code.

newStartHumanLoopResponse

Arguments

:: Int
$sel:httpStatus:StartHumanLoopResponse'

-> StartHumanLoopResponse

Create a value of StartHumanLoopResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:humanLoopArn:StartHumanLoopResponse', startHumanLoopResponse_humanLoopArn - The Amazon Resource Name (ARN) of the human loop.

$sel:httpStatus:StartHumanLoopResponse', startHumanLoopResponse_httpStatus - The response's http status code.

Response Lenses

startHumanLoopResponse_humanLoopArn :: Lens' StartHumanLoopResponse (Maybe Text)

The Amazon Resource Name (ARN) of the human loop.

startHumanLoopResponse_httpStatus :: Lens' StartHumanLoopResponse Int

The response's http status code.