Network.AWS.SageMakerA2IRuntime.Types

Description

Service Configuration

defaultService :: Service

API version 2019-11-07 of the Amazon Augmented AI Runtime SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The request isn't valid. Check the syntax and try again.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Your request has the same name as another active human loop but has different input data. You cannot start two human loops with the same name and different input data.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You exceeded your service quota. Service quotas, also referred to as limits, are the maximum number of service resources or operations for your AWS account. For a list of Amazon A2I service quotes, see Amazon Augmented AI Service Quotes. Delete some resources or request an increase in your service quota. You can request a quota increase using Service Quotas or the AWS Support Center. To request an increase, see AWS Service Quotas in the AWS General Reference.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

You exceeded the maximum number of requests.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

We couldn't process your request because of an issue with the server. Try again later.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

We couldn't find the requested resource. Check that your resources exists and were created in the same AWS Region as your request, and try your request again.

ContentClassifier

newtype ContentClassifier

Constructors

ContentClassifier'

Fields

fromContentClassifier :: Text

Bundled Patterns

pattern ContentClassifier_FreeOfAdultContent :: ContentClassifier

pattern ContentClassifier_FreeOfPersonallyIdentifiableInformation :: ContentClassifier

HumanLoopStatus

newtype HumanLoopStatus

Constructors

HumanLoopStatus'

Fields

fromHumanLoopStatus :: Text

Bundled Patterns

pattern HumanLoopStatus_Completed :: HumanLoopStatus

pattern HumanLoopStatus_Failed :: HumanLoopStatus

pattern HumanLoopStatus_InProgress :: HumanLoopStatus

pattern HumanLoopStatus_Stopped :: HumanLoopStatus

pattern HumanLoopStatus_Stopping :: HumanLoopStatus

SortOrder

newtype SortOrder

Constructors

SortOrder'

Fields

fromSortOrder :: Text

Bundled Patterns

pattern SortOrder_Ascending :: SortOrder

pattern SortOrder_Descending :: SortOrder

HumanLoopDataAttributes

data HumanLoopDataAttributes

Attributes of the data specified by the customer. Use these to describe the data to be labeled.

See: newHumanLoopDataAttributes smart constructor.

Constructors

HumanLoopDataAttributes'

Fields

contentClassifiers :: [ContentClassifier]

Declares that your content is free of personally identifiable information or adult content.

Amazon SageMaker can restrict the Amazon Mechanical Turk workers who can view your task based on this information.

newHumanLoopDataAttributes :: HumanLoopDataAttributes

Create a value of HumanLoopDataAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentClassifiers:HumanLoopDataAttributes', humanLoopDataAttributes_contentClassifiers - Declares that your content is free of personally identifiable information or adult content.

Amazon SageMaker can restrict the Amazon Mechanical Turk workers who can view your task based on this information.

humanLoopDataAttributes_contentClassifiers :: Lens' HumanLoopDataAttributes [ContentClassifier]

Declares that your content is free of personally identifiable information or adult content.

Amazon SageMaker can restrict the Amazon Mechanical Turk workers who can view your task based on this information.

HumanLoopInput

data HumanLoopInput

An object containing the human loop input in JSON format.

See: newHumanLoopInput smart constructor.

Constructors

HumanLoopInput'

Fields

inputContent :: Text

Serialized input from the human loop. The input must be a string representation of a file in JSON format.

newHumanLoopInput

Arguments

:: Text
$sel:inputContent:HumanLoopInput'

-> HumanLoopInput

Create a value of HumanLoopInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputContent:HumanLoopInput', humanLoopInput_inputContent - Serialized input from the human loop. The input must be a string representation of a file in JSON format.

humanLoopInput_inputContent :: Lens' HumanLoopInput Text

Serialized input from the human loop. The input must be a string representation of a file in JSON format.

HumanLoopOutput

data HumanLoopOutput

Information about where the human output will be stored.

See: newHumanLoopOutput smart constructor.

Constructors

HumanLoopOutput'

Fields

outputS3Uri :: Text

The location of the Amazon S3 object where Amazon Augmented AI stores your human loop output.

newHumanLoopOutput

Arguments

:: Text
$sel:outputS3Uri:HumanLoopOutput'

-> HumanLoopOutput

Create a value of HumanLoopOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outputS3Uri:HumanLoopOutput', humanLoopOutput_outputS3Uri - The location of the Amazon S3 object where Amazon Augmented AI stores your human loop output.

humanLoopOutput_outputS3Uri :: Lens' HumanLoopOutput Text

The location of the Amazon S3 object where Amazon Augmented AI stores your human loop output.

HumanLoopSummary

data HumanLoopSummary

Summary information about the human loop.

See: newHumanLoopSummary smart constructor.

Constructors

HumanLoopSummary'

Fields

creationTime :: Maybe POSIX

When Amazon Augmented AI created the human loop.

failureReason :: Maybe Text

The reason why the human loop failed. A failure reason is returned when the status of the human loop is Failed.

humanLoopStatus :: Maybe HumanLoopStatus

The status of the human loop.

humanLoopName :: Maybe Text

The name of the human loop.

flowDefinitionArn :: Maybe Text

The Amazon Resource Name (ARN) of the flow definition used to configure the human loop.

newHumanLoopSummary :: HumanLoopSummary

Create a value of HumanLoopSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:HumanLoopSummary', humanLoopSummary_creationTime - When Amazon Augmented AI created the human loop.

$sel:failureReason:HumanLoopSummary', humanLoopSummary_failureReason - The reason why the human loop failed. A failure reason is returned when the status of the human loop is Failed.

$sel:humanLoopStatus:HumanLoopSummary', humanLoopSummary_humanLoopStatus - The status of the human loop.

$sel:humanLoopName:HumanLoopSummary', humanLoopSummary_humanLoopName - The name of the human loop.

$sel:flowDefinitionArn:HumanLoopSummary', humanLoopSummary_flowDefinitionArn - The Amazon Resource Name (ARN) of the flow definition used to configure the human loop.

humanLoopSummary_creationTime :: Lens' HumanLoopSummary (Maybe UTCTime)

When Amazon Augmented AI created the human loop.

humanLoopSummary_failureReason :: Lens' HumanLoopSummary (Maybe Text)

The reason why the human loop failed. A failure reason is returned when the status of the human loop is Failed.

humanLoopSummary_humanLoopStatus :: Lens' HumanLoopSummary (Maybe HumanLoopStatus)

The status of the human loop.

humanLoopSummary_humanLoopName :: Lens' HumanLoopSummary (Maybe Text)

The name of the human loop.

humanLoopSummary_flowDefinitionArn :: Lens' HumanLoopSummary (Maybe Text)

The Amazon Resource Name (ARN) of the flow definition used to configure the human loop.