Network.AWS.SageMakerEdge.GetDeviceRegistration

Description

Use to check if a device is registered with SageMaker Edge Manager.

Creating a Request

data GetDeviceRegistration

See: newGetDeviceRegistration smart constructor.

Constructors

GetDeviceRegistration'

Fields

deviceName :: Text

The unique name of the device you want to get the registration status from.

deviceFleetName :: Text

The name of the fleet that the device belongs to.

newGetDeviceRegistration

Arguments

:: Text
$sel:deviceName:GetDeviceRegistration'

-> Text
$sel:deviceFleetName:GetDeviceRegistration'

-> GetDeviceRegistration

Create a value of GetDeviceRegistration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceName:GetDeviceRegistration', getDeviceRegistration_deviceName - The unique name of the device you want to get the registration status from.

$sel:deviceFleetName:GetDeviceRegistration', getDeviceRegistration_deviceFleetName - The name of the fleet that the device belongs to.

Request Lenses

getDeviceRegistration_deviceName :: Lens' GetDeviceRegistration Text

The unique name of the device you want to get the registration status from.

getDeviceRegistration_deviceFleetName :: Lens' GetDeviceRegistration Text

The name of the fleet that the device belongs to.

Destructuring the Response

data GetDeviceRegistrationResponse

See: newGetDeviceRegistrationResponse smart constructor.

Constructors

GetDeviceRegistrationResponse'

Fields

cacheTTL :: Maybe Text

The amount of time, in seconds, that the registration status is stored on the device’s cache before it is refreshed.

deviceRegistration :: Maybe Text

Describes if the device is currently registered with SageMaker Edge Manager.

httpStatus :: Int

The response's http status code.

newGetDeviceRegistrationResponse

Arguments

:: Int
$sel:httpStatus:GetDeviceRegistrationResponse'

-> GetDeviceRegistrationResponse

Create a value of GetDeviceRegistrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheTTL:GetDeviceRegistrationResponse', getDeviceRegistrationResponse_cacheTTL - The amount of time, in seconds, that the registration status is stored on the device’s cache before it is refreshed.

$sel:deviceRegistration:GetDeviceRegistrationResponse', getDeviceRegistrationResponse_deviceRegistration - Describes if the device is currently registered with SageMaker Edge Manager.

$sel:httpStatus:GetDeviceRegistrationResponse', getDeviceRegistrationResponse_httpStatus - The response's http status code.

Response Lenses

getDeviceRegistrationResponse_cacheTTL :: Lens' GetDeviceRegistrationResponse (Maybe Text)

The amount of time, in seconds, that the registration status is stored on the device’s cache before it is refreshed.

getDeviceRegistrationResponse_deviceRegistration :: Lens' GetDeviceRegistrationResponse (Maybe Text)

Describes if the device is currently registered with SageMaker Edge Manager.

getDeviceRegistrationResponse_httpStatus :: Lens' GetDeviceRegistrationResponse Int

The response's http status code.