Network.AWS.SageMakerEdge.Types

Description

Service Configuration

defaultService :: Service

API version 2020-09-23 of the Amazon Sagemaker Edge Manager SDK configuration.

Errors

_InternalServiceException :: AsError a => Getting (First ServiceError) a ServiceError

An internal failure occurred. Try your request again. If the problem persists, contact AWS customer support.

EdgeMetric

data EdgeMetric

Information required for edge device metrics.

See: newEdgeMetric smart constructor.

Constructors

EdgeMetric'

Fields

dimension :: Maybe Text

The dimension of metrics published.

metricName :: Maybe Text

Returns the name of the metric.

value :: Maybe Double

Returns the value of the metric.

timestamp :: Maybe POSIX

Timestamp of when the metric was requested.

newEdgeMetric :: EdgeMetric

Create a value of EdgeMetric with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimension:EdgeMetric', edgeMetric_dimension - The dimension of metrics published.

$sel:metricName:EdgeMetric', edgeMetric_metricName - Returns the name of the metric.

$sel:value:EdgeMetric', edgeMetric_value - Returns the value of the metric.

$sel:timestamp:EdgeMetric', edgeMetric_timestamp - Timestamp of when the metric was requested.

edgeMetric_dimension :: Lens' EdgeMetric (Maybe Text)

The dimension of metrics published.

edgeMetric_metricName :: Lens' EdgeMetric (Maybe Text)

Returns the name of the metric.

edgeMetric_value :: Lens' EdgeMetric (Maybe Double)

Returns the value of the metric.

edgeMetric_timestamp :: Lens' EdgeMetric (Maybe UTCTime)

Timestamp of when the metric was requested.

Model

data Model

Information about a model deployed on an edge device that is registered with SageMaker Edge Manager.

See: newModel smart constructor.

Constructors

Model'

Fields

modelName :: Maybe Text

The name of the model.

modelMetrics :: Maybe [EdgeMetric]

Information required for model metrics.

modelVersion :: Maybe Text

The version of the model.

latestInference :: Maybe POSIX

The timestamp of the last inference that was made.

latestSampleTime :: Maybe POSIX

The timestamp of the last data sample taken.

newModel :: Model

Create a value of Model with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelName:Model', model_modelName - The name of the model.

$sel:modelMetrics:Model', model_modelMetrics - Information required for model metrics.

$sel:modelVersion:Model', model_modelVersion - The version of the model.

$sel:latestInference:Model', model_latestInference - The timestamp of the last inference that was made.

$sel:latestSampleTime:Model', model_latestSampleTime - The timestamp of the last data sample taken.

model_modelName :: Lens' Model (Maybe Text)

The name of the model.

model_modelMetrics :: Lens' Model (Maybe [EdgeMetric])

Information required for model metrics.

model_modelVersion :: Lens' Model (Maybe Text)

The version of the model.

model_latestInference :: Lens' Model (Maybe UTCTime)

The timestamp of the last inference that was made.

model_latestSampleTime :: Lens' Model (Maybe UTCTime)

The timestamp of the last data sample taken.