Network.AWS.SageMakerFeatureStoreRuntime.Lens

Description

Operations

PutRecord

putRecord_featureGroupName :: Lens' PutRecord Text

The name of the feature group that you want to insert the record into.

putRecord_record :: Lens' PutRecord (NonEmpty FeatureValue)

List of FeatureValues to be inserted. This will be a full over-write. If you only want to update few of the feature values, do the following:

Use GetRecord to retrieve the latest record.

Update the record returned from GetRecord.

Use PutRecord to update feature values.

DeleteRecord

deleteRecord_featureGroupName :: Lens' DeleteRecord Text

The name of the feature group to delete the record from.

deleteRecord_recordIdentifierValueAsString :: Lens' DeleteRecord Text

The value for the RecordIdentifier that uniquely identifies the record, in string format.

deleteRecord_eventTime :: Lens' DeleteRecord Text

Timestamp indicating when the deletion event occurred. EventTime can be used to query data at a certain point in time.

BatchGetRecord

batchGetRecord_identifiers :: Lens' BatchGetRecord (NonEmpty BatchGetRecordIdentifier)

A list of FeatureGroup names, with their corresponding RecordIdentifier value, and Feature name that have been requested to be retrieved in batch.

batchGetRecordResponse_httpStatus :: Lens' BatchGetRecordResponse Int

The response's http status code.

batchGetRecordResponse_records :: Lens' BatchGetRecordResponse [BatchGetRecordResultDetail]

A list of Records you requested to be retrieved in batch.

batchGetRecordResponse_errors :: Lens' BatchGetRecordResponse [BatchGetRecordError]

A list of errors that have occured when retrieving a batch of Records.

batchGetRecordResponse_unprocessedIdentifiers :: Lens' BatchGetRecordResponse [BatchGetRecordIdentifier]

A unprocessed list of FeatureGroup names, with their corresponding RecordIdentifier value, and Feature name.

GetRecord

getRecord_featureNames :: Lens' GetRecord (Maybe (NonEmpty Text))

List of names of Features to be retrieved. If not specified, the latest value for all the Features are returned.

getRecord_featureGroupName :: Lens' GetRecord Text

The name of the feature group in which you want to put the records.

getRecord_recordIdentifierValueAsString :: Lens' GetRecord Text

The value that corresponds to RecordIdentifier type and uniquely identifies the record in the FeatureGroup.

getRecordResponse_record :: Lens' GetRecordResponse (Maybe (NonEmpty FeatureValue))

The record you requested. A list of FeatureValues.

getRecordResponse_httpStatus :: Lens' GetRecordResponse Int

The response's http status code.

Types

BatchGetRecordError

batchGetRecordError_featureGroupName :: Lens' BatchGetRecordError Text

The name of the feature group that the record belongs to.

batchGetRecordError_recordIdentifierValueAsString :: Lens' BatchGetRecordError Text

The value for the RecordIdentifier in string format of a Record from a FeatureGroup that is causing an error when attempting to be retrieved.

batchGetRecordError_errorCode :: Lens' BatchGetRecordError Text

The error code of an error that has occured when attempting to retrieve a batch of Records. For more information on errors, see Errors.

batchGetRecordError_errorMessage :: Lens' BatchGetRecordError Text

The error message of an error that has occured when attempting to retrieve a record in the batch.

BatchGetRecordIdentifier

batchGetRecordIdentifier_featureNames :: Lens' BatchGetRecordIdentifier (Maybe (NonEmpty Text))

List of names of Features to be retrieved. If not specified, the latest value for all the Features are returned.

batchGetRecordIdentifier_featureGroupName :: Lens' BatchGetRecordIdentifier Text

A FeatureGroupName containing Records you are retrieving in a batch.

batchGetRecordIdentifier_recordIdentifiersValueAsString :: Lens' BatchGetRecordIdentifier (NonEmpty Text)

The value for a list of record identifiers in string format.

BatchGetRecordResultDetail

batchGetRecordResultDetail_featureGroupName :: Lens' BatchGetRecordResultDetail Text

The FeatureGroupName containing Records you retrieved in a batch.

batchGetRecordResultDetail_recordIdentifierValueAsString :: Lens' BatchGetRecordResultDetail Text

The value of the record identifer in string format.

batchGetRecordResultDetail_record :: Lens' BatchGetRecordResultDetail (NonEmpty FeatureValue)

The Record retrieved.

FeatureValue

featureValue_featureName :: Lens' FeatureValue Text

The name of a feature that a feature value corresponds to.

featureValue_valueAsString :: Lens' FeatureValue Text

The value associated with a feature, in string format. Note that features types can be String, Integral, or Fractional. This value represents all three types as a string.