Network.AWS.SageMakerFeatureStoreRuntime.PutRecord

Description

Used for data ingestion into the FeatureStore. The PutRecord API writes to both the OnlineStore and OfflineStore. If the record is the latest record for the recordIdentifier, the record is written to both the OnlineStore and OfflineStore. If the record is a historic record, it is written only to the OfflineStore.

Creating a Request

data PutRecord

See: newPutRecord smart constructor.

Constructors

PutRecord'

Fields

featureGroupName :: Text

The name of the feature group that you want to insert the record into.

record :: NonEmpty FeatureValue

List of FeatureValues to be inserted. This will be a full over-write. If you only want to update few of the feature values, do the following:

Use GetRecord to retrieve the latest record.

Update the record returned from GetRecord.

Use PutRecord to update feature values.

newPutRecord

Arguments

:: Text
$sel:featureGroupName:PutRecord'

-> NonEmpty FeatureValue
$sel:record:PutRecord'

-> PutRecord

Create a value of PutRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureGroupName:PutRecord', putRecord_featureGroupName - The name of the feature group that you want to insert the record into.

$sel:record:PutRecord', putRecord_record - List of FeatureValues to be inserted. This will be a full over-write. If you only want to update few of the feature values, do the following:

Use GetRecord to retrieve the latest record.

Update the record returned from GetRecord.

Use PutRecord to update feature values.

Request Lenses

putRecord_featureGroupName :: Lens' PutRecord Text

The name of the feature group that you want to insert the record into.

putRecord_record :: Lens' PutRecord (NonEmpty FeatureValue)

List of FeatureValues to be inserted. This will be a full over-write. If you only want to update few of the feature values, do the following:

Use GetRecord to retrieve the latest record.

Update the record returned from GetRecord.

Use PutRecord to update feature values.

Destructuring the Response

data PutRecordResponse

See: newPutRecordResponse smart constructor.

Constructors

PutRecordResponse'

newPutRecordResponse :: PutRecordResponse

Create a value of PutRecordResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.