Network.AWS.SageMakerFeatureStoreRuntime.Types

Description

Service Configuration

defaultService :: Service

API version 2020-07-01 of the Amazon SageMaker Feature Store Runtime SDK configuration.

Errors

_AccessForbidden :: AsError a => Getting (First ServiceError) a ServiceError

You do not have permission to perform an action.

_ServiceUnavailable :: AsError a => Getting (First ServiceError) a ServiceError

The service is currently unavailable.

_InternalFailure :: AsError a => Getting (First ServiceError) a ServiceError

An internal failure occurred. Try your request again. If the problem persists, contact AWS customer support.

_ValidationError :: AsError a => Getting (First ServiceError) a ServiceError

There was an error validating your request.

_ResourceNotFound :: AsError a => Getting (First ServiceError) a ServiceError

A resource that is required to perform an action was not found.

BatchGetRecordError

data BatchGetRecordError

The error that has occurred when attempting to retrieve a batch of Records.

See: newBatchGetRecordError smart constructor.

Constructors

BatchGetRecordError'

Fields

featureGroupName :: Text

The name of the feature group that the record belongs to.

recordIdentifierValueAsString :: Text

The value for the RecordIdentifier in string format of a Record from a FeatureGroup that is causing an error when attempting to be retrieved.

errorCode :: Text

The error code of an error that has occured when attempting to retrieve a batch of Records. For more information on errors, see Errors.

errorMessage :: Text

The error message of an error that has occured when attempting to retrieve a record in the batch.

newBatchGetRecordError

Arguments

:: Text
$sel:featureGroupName:BatchGetRecordError'

-> Text
$sel:recordIdentifierValueAsString:BatchGetRecordError'

-> Text
$sel:errorCode:BatchGetRecordError'

-> Text
$sel:errorMessage:BatchGetRecordError'

-> BatchGetRecordError

Create a value of BatchGetRecordError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureGroupName:BatchGetRecordError', batchGetRecordError_featureGroupName - The name of the feature group that the record belongs to.

$sel:recordIdentifierValueAsString:BatchGetRecordError', batchGetRecordError_recordIdentifierValueAsString - The value for the RecordIdentifier in string format of a Record from a FeatureGroup that is causing an error when attempting to be retrieved.

$sel:errorCode:BatchGetRecordError', batchGetRecordError_errorCode - The error code of an error that has occured when attempting to retrieve a batch of Records. For more information on errors, see Errors.

$sel:errorMessage:BatchGetRecordError', batchGetRecordError_errorMessage - The error message of an error that has occured when attempting to retrieve a record in the batch.

batchGetRecordError_featureGroupName :: Lens' BatchGetRecordError Text

The name of the feature group that the record belongs to.

batchGetRecordError_recordIdentifierValueAsString :: Lens' BatchGetRecordError Text

The value for the RecordIdentifier in string format of a Record from a FeatureGroup that is causing an error when attempting to be retrieved.

batchGetRecordError_errorCode :: Lens' BatchGetRecordError Text

The error code of an error that has occured when attempting to retrieve a batch of Records. For more information on errors, see Errors.

batchGetRecordError_errorMessage :: Lens' BatchGetRecordError Text

The error message of an error that has occured when attempting to retrieve a record in the batch.

BatchGetRecordIdentifier

data BatchGetRecordIdentifier

The identifier that identifies the batch of Records you are retrieving in a batch.

See: newBatchGetRecordIdentifier smart constructor.

Constructors

BatchGetRecordIdentifier'

Fields

featureNames :: Maybe (NonEmpty Text)

List of names of Features to be retrieved. If not specified, the latest value for all the Features are returned.

featureGroupName :: Text

A FeatureGroupName containing Records you are retrieving in a batch.

recordIdentifiersValueAsString :: NonEmpty Text

The value for a list of record identifiers in string format.

newBatchGetRecordIdentifier

Arguments

:: Text
$sel:featureGroupName:BatchGetRecordIdentifier'

-> NonEmpty Text
$sel:recordIdentifiersValueAsString:BatchGetRecordIdentifier'

-> BatchGetRecordIdentifier

Create a value of BatchGetRecordIdentifier with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureNames:BatchGetRecordIdentifier', batchGetRecordIdentifier_featureNames - List of names of Features to be retrieved. If not specified, the latest value for all the Features are returned.

$sel:featureGroupName:BatchGetRecordIdentifier', batchGetRecordIdentifier_featureGroupName - A FeatureGroupName containing Records you are retrieving in a batch.

$sel:recordIdentifiersValueAsString:BatchGetRecordIdentifier', batchGetRecordIdentifier_recordIdentifiersValueAsString - The value for a list of record identifiers in string format.

batchGetRecordIdentifier_featureNames :: Lens' BatchGetRecordIdentifier (Maybe (NonEmpty Text))

List of names of Features to be retrieved. If not specified, the latest value for all the Features are returned.

batchGetRecordIdentifier_featureGroupName :: Lens' BatchGetRecordIdentifier Text

A FeatureGroupName containing Records you are retrieving in a batch.

batchGetRecordIdentifier_recordIdentifiersValueAsString :: Lens' BatchGetRecordIdentifier (NonEmpty Text)

The value for a list of record identifiers in string format.

BatchGetRecordResultDetail

data BatchGetRecordResultDetail

The output of Records that have been retrieved in a batch.

See: newBatchGetRecordResultDetail smart constructor.

Constructors

BatchGetRecordResultDetail'

Fields

featureGroupName :: Text

The FeatureGroupName containing Records you retrieved in a batch.

recordIdentifierValueAsString :: Text

The value of the record identifer in string format.

record :: NonEmpty FeatureValue

The Record retrieved.

newBatchGetRecordResultDetail

Arguments

:: Text
$sel:featureGroupName:BatchGetRecordResultDetail'

-> Text
$sel:recordIdentifierValueAsString:BatchGetRecordResultDetail'

-> NonEmpty FeatureValue
$sel:record:BatchGetRecordResultDetail'

-> BatchGetRecordResultDetail

Create a value of BatchGetRecordResultDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureGroupName:BatchGetRecordResultDetail', batchGetRecordResultDetail_featureGroupName - The FeatureGroupName containing Records you retrieved in a batch.

$sel:recordIdentifierValueAsString:BatchGetRecordResultDetail', batchGetRecordResultDetail_recordIdentifierValueAsString - The value of the record identifer in string format.

$sel:record:BatchGetRecordResultDetail', batchGetRecordResultDetail_record - The Record retrieved.

batchGetRecordResultDetail_featureGroupName :: Lens' BatchGetRecordResultDetail Text

The FeatureGroupName containing Records you retrieved in a batch.

batchGetRecordResultDetail_recordIdentifierValueAsString :: Lens' BatchGetRecordResultDetail Text

The value of the record identifer in string format.

batchGetRecordResultDetail_record :: Lens' BatchGetRecordResultDetail (NonEmpty FeatureValue)

The Record retrieved.

FeatureValue

data FeatureValue

The value associated with a feature.

See: newFeatureValue smart constructor.

Constructors

FeatureValue'

Fields

featureName :: Text

The name of a feature that a feature value corresponds to.

valueAsString :: Text

The value associated with a feature, in string format. Note that features types can be String, Integral, or Fractional. This value represents all three types as a string.

newFeatureValue

Arguments

:: Text
$sel:featureName:FeatureValue'

-> Text
$sel:valueAsString:FeatureValue'

-> FeatureValue

Create a value of FeatureValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureName:FeatureValue', featureValue_featureName - The name of a feature that a feature value corresponds to.

$sel:valueAsString:FeatureValue', featureValue_valueAsString - The value associated with a feature, in string format. Note that features types can be String, Integral, or Fractional. This value represents all three types as a string.

featureValue_featureName :: Lens' FeatureValue Text

The name of a feature that a feature value corresponds to.

featureValue_valueAsString :: Lens' FeatureValue Text

The value associated with a feature, in string format. Note that features types can be String, Integral, or Fractional. This value represents all three types as a string.