Network.AWS.SageMakerRuntime.InvokeEndpointAsync

Description

After you deploy a model into production using Amazon SageMaker hosting services, your client applications use this API to get inferences from the model hosted at the specified endpoint in an asynchronous manner.

Inference requests sent to this API are enqueued for asynchronous processing. The processing of the inference request may or may not complete before the you receive a response from this API. The response from this API will not contain the result of the inference request but contain information about where you can locate it.

Amazon SageMaker strips all POST headers except those supported by the API. Amazon SageMaker might add additional headers. You should not rely on the behavior of headers outside those enumerated in the request syntax.

Calls to InvokeEndpointAsync are authenticated by using AWS Signature Version 4. For information, see Authenticating Requests (AWS Signature Version 4) in the Amazon S3 API Reference.

Creating a Request

data InvokeEndpointAsync

See: newInvokeEndpointAsync smart constructor.

Constructors

InvokeEndpointAsync'

Fields

contentType :: Maybe Text

The MIME type of the input data in the request body.

accept :: Maybe Text

The desired MIME type of the inference in the response.

customAttributes :: Maybe (Sensitive Text)

Provides additional information about a request for an inference submitted to a model hosted at an Amazon SageMaker endpoint. The information is an opaque value that is forwarded verbatim. You could use this value, for example, to provide an ID that you can use to track a request or to provide other metadata that a service endpoint was programmed to process. The value must consist of no more than 1024 visible US-ASCII characters as specified in Section 3.3.6. Field Value Components of the Hypertext Transfer Protocol (HTTP/1.1).

The code in your model is responsible for setting or updating any custom attributes in the response. If your code does not set this value in the response, an empty value is returned. For example, if a custom attribute represents the trace ID, your model can prepend the custom attribute with Trace ID: in your post-processing function.

This feature is currently supported in the AWS SDKs but not in the Amazon SageMaker Python SDK.

requestTTLSeconds :: Maybe Natural

Maximum age in seconds a request can be in the queue before it is marked as expired.

inferenceId :: Maybe Text

The identifier for the inference request. Amazon SageMaker will generate an identifier for you if none is specified.

endpointName :: Text

The name of the endpoint that you specified when you created the endpoint using the CreateEndpoint API.

inputLocation :: Text

The Amazon S3 URI where the inference request payload is stored.

newInvokeEndpointAsync

Arguments

:: Text
$sel:endpointName:InvokeEndpointAsync'

-> Text
$sel:inputLocation:InvokeEndpointAsync'

-> InvokeEndpointAsync

Create a value of InvokeEndpointAsync with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentType:InvokeEndpointAsync', invokeEndpointAsync_contentType - The MIME type of the input data in the request body.

$sel:accept:InvokeEndpointAsync', invokeEndpointAsync_accept - The desired MIME type of the inference in the response.

$sel:customAttributes:InvokeEndpointAsync', invokeEndpointAsync_customAttributes - Provides additional information about a request for an inference submitted to a model hosted at an Amazon SageMaker endpoint. The information is an opaque value that is forwarded verbatim. You could use this value, for example, to provide an ID that you can use to track a request or to provide other metadata that a service endpoint was programmed to process. The value must consist of no more than 1024 visible US-ASCII characters as specified in Section 3.3.6. Field Value Components of the Hypertext Transfer Protocol (HTTP/1.1).

The code in your model is responsible for setting or updating any custom attributes in the response. If your code does not set this value in the response, an empty value is returned. For example, if a custom attribute represents the trace ID, your model can prepend the custom attribute with Trace ID: in your post-processing function.

This feature is currently supported in the AWS SDKs but not in the Amazon SageMaker Python SDK.

$sel:requestTTLSeconds:InvokeEndpointAsync', invokeEndpointAsync_requestTTLSeconds - Maximum age in seconds a request can be in the queue before it is marked as expired.

$sel:inferenceId:InvokeEndpointAsync', invokeEndpointAsync_inferenceId - The identifier for the inference request. Amazon SageMaker will generate an identifier for you if none is specified.

$sel:endpointName:InvokeEndpointAsync', invokeEndpointAsync_endpointName - The name of the endpoint that you specified when you created the endpoint using the CreateEndpoint API.

$sel:inputLocation:InvokeEndpointAsync', invokeEndpointAsync_inputLocation - The Amazon S3 URI where the inference request payload is stored.

Request Lenses

invokeEndpointAsync_contentType :: Lens' InvokeEndpointAsync (Maybe Text)

The MIME type of the input data in the request body.

invokeEndpointAsync_accept :: Lens' InvokeEndpointAsync (Maybe Text)

The desired MIME type of the inference in the response.

invokeEndpointAsync_customAttributes :: Lens' InvokeEndpointAsync (Maybe Text)

Provides additional information about a request for an inference submitted to a model hosted at an Amazon SageMaker endpoint. The information is an opaque value that is forwarded verbatim. You could use this value, for example, to provide an ID that you can use to track a request or to provide other metadata that a service endpoint was programmed to process. The value must consist of no more than 1024 visible US-ASCII characters as specified in Section 3.3.6. Field Value Components of the Hypertext Transfer Protocol (HTTP/1.1).

The code in your model is responsible for setting or updating any custom attributes in the response. If your code does not set this value in the response, an empty value is returned. For example, if a custom attribute represents the trace ID, your model can prepend the custom attribute with Trace ID: in your post-processing function.

This feature is currently supported in the AWS SDKs but not in the Amazon SageMaker Python SDK.

invokeEndpointAsync_requestTTLSeconds :: Lens' InvokeEndpointAsync (Maybe Natural)

Maximum age in seconds a request can be in the queue before it is marked as expired.

invokeEndpointAsync_inferenceId :: Lens' InvokeEndpointAsync (Maybe Text)

The identifier for the inference request. Amazon SageMaker will generate an identifier for you if none is specified.

invokeEndpointAsync_endpointName :: Lens' InvokeEndpointAsync Text

The name of the endpoint that you specified when you created the endpoint using the CreateEndpoint API.

invokeEndpointAsync_inputLocation :: Lens' InvokeEndpointAsync Text

The Amazon S3 URI where the inference request payload is stored.

Destructuring the Response

data InvokeEndpointAsyncResponse

See: newInvokeEndpointAsyncResponse smart constructor.

Constructors

InvokeEndpointAsyncResponse'

Fields

inferenceId :: Maybe Text

Identifier for an inference request. This will be the same as the InferenceId specified in the input. Amazon SageMaker will generate an identifier for you if you do not specify one.

outputLocation :: Maybe Text

The Amazon S3 URI where the inference response payload is stored.

httpStatus :: Int

The response's http status code.

newInvokeEndpointAsyncResponse

Arguments

:: Int
$sel:httpStatus:InvokeEndpointAsyncResponse'

-> InvokeEndpointAsyncResponse

Create a value of InvokeEndpointAsyncResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inferenceId:InvokeEndpointAsync', invokeEndpointAsyncResponse_inferenceId - Identifier for an inference request. This will be the same as the InferenceId specified in the input. Amazon SageMaker will generate an identifier for you if you do not specify one.

$sel:outputLocation:InvokeEndpointAsyncResponse', invokeEndpointAsyncResponse_outputLocation - The Amazon S3 URI where the inference response payload is stored.

$sel:httpStatus:InvokeEndpointAsyncResponse', invokeEndpointAsyncResponse_httpStatus - The response's http status code.

Response Lenses

invokeEndpointAsyncResponse_inferenceId :: Lens' InvokeEndpointAsyncResponse (Maybe Text)

Identifier for an inference request. This will be the same as the InferenceId specified in the input. Amazon SageMaker will generate an identifier for you if you do not specify one.

invokeEndpointAsyncResponse_outputLocation :: Lens' InvokeEndpointAsyncResponse (Maybe Text)

The Amazon S3 URI where the inference response payload is stored.

invokeEndpointAsyncResponse_httpStatus :: Lens' InvokeEndpointAsyncResponse Int

The response's http status code.