Network.AWS.SageMakerRuntime.Lens

Description

Operations

InvokeEndpoint

invokeEndpoint_targetContainerHostname :: Lens' InvokeEndpoint (Maybe Text)

If the endpoint hosts multiple containers and is configured to use direct invocation, this parameter specifies the host name of the container to invoke.

invokeEndpoint_contentType :: Lens' InvokeEndpoint (Maybe Text)

The MIME type of the input data in the request body.

invokeEndpoint_targetModel :: Lens' InvokeEndpoint (Maybe Text)

The model to request for inference when invoking a multi-model endpoint.

invokeEndpoint_accept :: Lens' InvokeEndpoint (Maybe Text)

The desired MIME type of the inference in the response.

invokeEndpoint_customAttributes :: Lens' InvokeEndpoint (Maybe Text)

Provides additional information about a request for an inference submitted to a model hosted at an Amazon SageMaker endpoint. The information is an opaque value that is forwarded verbatim. You could use this value, for example, to provide an ID that you can use to track a request or to provide other metadata that a service endpoint was programmed to process. The value must consist of no more than 1024 visible US-ASCII characters as specified in Section 3.3.6. Field Value Components of the Hypertext Transfer Protocol (HTTP/1.1).

The code in your model is responsible for setting or updating any custom attributes in the response. If your code does not set this value in the response, an empty value is returned. For example, if a custom attribute represents the trace ID, your model can prepend the custom attribute with Trace ID: in your post-processing function.

This feature is currently supported in the AWS SDKs but not in the Amazon SageMaker Python SDK.

invokeEndpoint_targetVariant :: Lens' InvokeEndpoint (Maybe Text)

Specify the production variant to send the inference request to when invoking an endpoint that is running two or more variants. Note that this parameter overrides the default behavior for the endpoint, which is to distribute the invocation traffic based on the variant weights.

For information about how to use variant targeting to perform a/b testing, see Test models in production

invokeEndpoint_inferenceId :: Lens' InvokeEndpoint (Maybe Text)

If you provide a value, it is added to the captured data when you enable data capture on the endpoint. For information about data capture, see Capture Data.

invokeEndpoint_endpointName :: Lens' InvokeEndpoint Text

The name of the endpoint that you specified when you created the endpoint using the CreateEndpoint API.

invokeEndpoint_body :: Lens' InvokeEndpoint Text

Provides input data, in the format specified in the ContentType request header. Amazon SageMaker passes all of the data in the body to the model.

For information about the format of the request body, see Common Data Formats-Inference.

invokeEndpointResponse_contentType :: Lens' InvokeEndpointResponse (Maybe Text)

The MIME type of the inference returned in the response body.

invokeEndpointResponse_invokedProductionVariant :: Lens' InvokeEndpointResponse (Maybe Text)

Identifies the production variant that was invoked.

invokeEndpointResponse_customAttributes :: Lens' InvokeEndpointResponse (Maybe Text)

Provides additional information in the response about the inference returned by a model hosted at an Amazon SageMaker endpoint. The information is an opaque value that is forwarded verbatim. You could use this value, for example, to return an ID received in the CustomAttributes header of a request or other metadata that a service endpoint was programmed to produce. The value must consist of no more than 1024 visible US-ASCII characters as specified in Section 3.3.6. Field Value Components of the Hypertext Transfer Protocol (HTTP/1.1). If the customer wants the custom attribute returned, the model must set the custom attribute to be included on the way back.

The code in your model is responsible for setting or updating any custom attributes in the response. If your code does not set this value in the response, an empty value is returned. For example, if a custom attribute represents the trace ID, your model can prepend the custom attribute with Trace ID: in your post-processing function.

This feature is currently supported in the AWS SDKs but not in the Amazon SageMaker Python SDK.

invokeEndpointResponse_httpStatus :: Lens' InvokeEndpointResponse Int

The response's http status code.

invokeEndpointResponse_body :: Lens' InvokeEndpointResponse Text

Includes the inference provided by the model.

For information about the format of the response body, see Common Data Formats-Inference.

InvokeEndpointAsync

invokeEndpointAsync_contentType :: Lens' InvokeEndpointAsync (Maybe Text)

The MIME type of the input data in the request body.

invokeEndpointAsync_accept :: Lens' InvokeEndpointAsync (Maybe Text)

The desired MIME type of the inference in the response.

invokeEndpointAsync_customAttributes :: Lens' InvokeEndpointAsync (Maybe Text)

Provides additional information about a request for an inference submitted to a model hosted at an Amazon SageMaker endpoint. The information is an opaque value that is forwarded verbatim. You could use this value, for example, to provide an ID that you can use to track a request or to provide other metadata that a service endpoint was programmed to process. The value must consist of no more than 1024 visible US-ASCII characters as specified in Section 3.3.6. Field Value Components of the Hypertext Transfer Protocol (HTTP/1.1).

The code in your model is responsible for setting or updating any custom attributes in the response. If your code does not set this value in the response, an empty value is returned. For example, if a custom attribute represents the trace ID, your model can prepend the custom attribute with Trace ID: in your post-processing function.

This feature is currently supported in the AWS SDKs but not in the Amazon SageMaker Python SDK.

invokeEndpointAsync_requestTTLSeconds :: Lens' InvokeEndpointAsync (Maybe Natural)

Maximum age in seconds a request can be in the queue before it is marked as expired.

invokeEndpointAsync_inferenceId :: Lens' InvokeEndpointAsync (Maybe Text)

The identifier for the inference request. Amazon SageMaker will generate an identifier for you if none is specified.

invokeEndpointAsync_endpointName :: Lens' InvokeEndpointAsync Text

The name of the endpoint that you specified when you created the endpoint using the CreateEndpoint API.

invokeEndpointAsync_inputLocation :: Lens' InvokeEndpointAsync Text

The Amazon S3 URI where the inference request payload is stored.

invokeEndpointAsyncResponse_inferenceId :: Lens' InvokeEndpointAsyncResponse (Maybe Text)

Identifier for an inference request. This will be the same as the InferenceId specified in the input. Amazon SageMaker will generate an identifier for you if you do not specify one.

invokeEndpointAsyncResponse_outputLocation :: Lens' InvokeEndpointAsyncResponse (Maybe Text)

The Amazon S3 URI where the inference response payload is stored.

invokeEndpointAsyncResponse_httpStatus :: Lens' InvokeEndpointAsyncResponse Int

The response's http status code.

Types