Network.AWS.SavingsPlans

Description

Derived from API version 2019-06-28 of the AWS service descriptions, licensed under Apache 2.0.

Savings Plans are a pricing model that offer significant savings on AWS usage (for example, on Amazon EC2 instances). You commit to a consistent amount of usage, in USD per hour, for a term of 1 or 3 years, and receive a lower price for that usage. For more information, see the AWS Savings Plans User Guide.

Service Configuration

defaultService :: Service

API version 2019-06-28 of the Amazon Savings Plans SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by SavingsPlans.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

One of the input parameters is not valid.

ServiceQuotaExceededException

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A service quota has been exceeded.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

An unexpected error occurred.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource was not found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The Amazon Resource Name (ARN) of the resource.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - Information about the tags.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

DescribeSavingsPlanRates

data DescribeSavingsPlanRates

See: newDescribeSavingsPlanRates smart constructor.

Constructors

DescribeSavingsPlanRates' (Maybe [SavingsPlanRateFilter]) (Maybe Text) (Maybe Natural) Text

newDescribeSavingsPlanRates

Arguments

:: Text
$sel:savingsPlanId:DescribeSavingsPlanRates'

-> DescribeSavingsPlanRates

Create a value of DescribeSavingsPlanRates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeSavingsPlanRates', describeSavingsPlanRates_filters - The filters.

$sel:nextToken:DescribeSavingsPlanRates', describeSavingsPlanRates_nextToken - The token for the next page of results.

$sel:maxResults:DescribeSavingsPlanRates', describeSavingsPlanRates_maxResults - The maximum number of results to return with a single call. To retrieve additional results, make another call with the returned token value.

$sel:savingsPlanId:DescribeSavingsPlanRates', describeSavingsPlanRates_savingsPlanId - The ID of the Savings Plan.

data DescribeSavingsPlanRatesResponse

See: newDescribeSavingsPlanRatesResponse smart constructor.

Constructors

DescribeSavingsPlanRatesResponse' (Maybe [SavingsPlanRate]) (Maybe Text) (Maybe Text) Int

newDescribeSavingsPlanRatesResponse

Arguments

:: Int
$sel:httpStatus:DescribeSavingsPlanRatesResponse'

-> DescribeSavingsPlanRatesResponse

Create a value of DescribeSavingsPlanRatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:searchResults:DescribeSavingsPlanRatesResponse', describeSavingsPlanRatesResponse_searchResults - Information about the Savings Plans rates.

$sel:savingsPlanId:DescribeSavingsPlanRates', describeSavingsPlanRatesResponse_savingsPlanId - The ID of the Savings Plan.

$sel:nextToken:DescribeSavingsPlanRates', describeSavingsPlanRatesResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeSavingsPlanRatesResponse', describeSavingsPlanRatesResponse_httpStatus - The response's http status code.

DeleteQueuedSavingsPlan

data DeleteQueuedSavingsPlan

See: newDeleteQueuedSavingsPlan smart constructor.

Constructors

DeleteQueuedSavingsPlan' Text

newDeleteQueuedSavingsPlan

Arguments

:: Text
$sel:savingsPlanId:DeleteQueuedSavingsPlan'

-> DeleteQueuedSavingsPlan

Create a value of DeleteQueuedSavingsPlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:savingsPlanId:DeleteQueuedSavingsPlan', deleteQueuedSavingsPlan_savingsPlanId - The ID of the Savings Plan.

data DeleteQueuedSavingsPlanResponse

See: newDeleteQueuedSavingsPlanResponse smart constructor.

Constructors

DeleteQueuedSavingsPlanResponse' Int

newDeleteQueuedSavingsPlanResponse

Arguments

:: Int
$sel:httpStatus:DeleteQueuedSavingsPlanResponse'

-> DeleteQueuedSavingsPlanResponse

Create a value of DeleteQueuedSavingsPlanResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteQueuedSavingsPlanResponse', deleteQueuedSavingsPlanResponse_httpStatus - The response's http status code.

CreateSavingsPlan

data CreateSavingsPlan

See: newCreateSavingsPlan smart constructor.

Constructors

CreateSavingsPlan' (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe (HashMap Text Text)) Text Text

newCreateSavingsPlan

Arguments

:: Text
$sel:savingsPlanOfferingId:CreateSavingsPlan'

-> Text
$sel:commitment:CreateSavingsPlan'

-> CreateSavingsPlan

Create a value of CreateSavingsPlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateSavingsPlan', createSavingsPlan_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:purchaseTime:CreateSavingsPlan', createSavingsPlan_purchaseTime - The time at which to purchase the Savings Plan, in UTC format (YYYY-MM-DDTHH:MM:SSZ).

$sel:upfrontPaymentAmount:CreateSavingsPlan', createSavingsPlan_upfrontPaymentAmount - The up-front payment amount. This is a whole number between 50 and 99 percent of the total value of the Savings Plan. This parameter is supported only if the payment option is Partial Upfront.

$sel:tags:CreateSavingsPlan', createSavingsPlan_tags - One or more tags.

$sel:savingsPlanOfferingId:CreateSavingsPlan', createSavingsPlan_savingsPlanOfferingId - The ID of the offering.

$sel:commitment:CreateSavingsPlan', createSavingsPlan_commitment - The hourly commitment, in USD. This is a value between 0.001 and 1 million. You cannot specify more than three digits after the decimal point.

data CreateSavingsPlanResponse

See: newCreateSavingsPlanResponse smart constructor.

Constructors

CreateSavingsPlanResponse' (Maybe Text) Int

newCreateSavingsPlanResponse

Arguments

:: Int
$sel:httpStatus:CreateSavingsPlanResponse'

-> CreateSavingsPlanResponse

Create a value of CreateSavingsPlanResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:savingsPlanId:CreateSavingsPlanResponse', createSavingsPlanResponse_savingsPlanId - The ID of the Savings Plan.

$sel:httpStatus:CreateSavingsPlanResponse', createSavingsPlanResponse_httpStatus - The response's http status code.

DescribeSavingsPlansOfferings

data DescribeSavingsPlansOfferings

See: newDescribeSavingsPlansOfferings smart constructor.

Constructors

DescribeSavingsPlansOfferings' (Maybe [Text]) (Maybe SavingsPlanProductType) (Maybe [SavingsPlanOfferingFilterElement]) (Maybe [Text]) (Maybe [CurrencyCode]) (Maybe Text) (Maybe [SavingsPlanPaymentOption]) (Maybe [Text]) (Maybe [Natural]) (Maybe [SavingsPlanType]) (Maybe [Text]) (Maybe [Text]) (Maybe Natural)

newDescribeSavingsPlansOfferings :: DescribeSavingsPlansOfferings

Create a value of DescribeSavingsPlansOfferings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceCodes:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_serviceCodes - The services.

$sel:productType:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_productType - The product type.

$sel:filters:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_filters - The filters.

$sel:offeringIds:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_offeringIds - The IDs of the offerings.

$sel:currencies:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_currencies - The currencies.

$sel:nextToken:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_nextToken - The token for the next page of results.

$sel:paymentOptions:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_paymentOptions - The payment options.

$sel:descriptions:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_descriptions - The descriptions.

$sel:durations:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_durations - The durations, in seconds.

$sel:planTypes:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_planTypes - The plan type.

$sel:usageTypes:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_usageTypes - The usage details of the line item in the billing report.

$sel:operations:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_operations - The specific AWS operation for the line item in the billing report.

$sel:maxResults:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_maxResults - The maximum number of results to return with a single call. To retrieve additional results, make another call with the returned token value.

data DescribeSavingsPlansOfferingsResponse

See: newDescribeSavingsPlansOfferingsResponse smart constructor.

Constructors

DescribeSavingsPlansOfferingsResponse' (Maybe [SavingsPlanOffering]) (Maybe Text) Int

newDescribeSavingsPlansOfferingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeSavingsPlansOfferingsResponse'

-> DescribeSavingsPlansOfferingsResponse

Create a value of DescribeSavingsPlansOfferingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:searchResults:DescribeSavingsPlansOfferingsResponse', describeSavingsPlansOfferingsResponse_searchResults - Information about the Savings Plans offerings.

$sel:nextToken:DescribeSavingsPlansOfferings', describeSavingsPlansOfferingsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeSavingsPlansOfferingsResponse', describeSavingsPlansOfferingsResponse_httpStatus - The response's http status code.

DescribeSavingsPlans

data DescribeSavingsPlans

See: newDescribeSavingsPlans smart constructor.

Constructors

DescribeSavingsPlans' (Maybe [SavingsPlanState]) (Maybe [Text]) (Maybe [SavingsPlanFilter]) (Maybe Text) (Maybe [Text]) (Maybe Natural)

newDescribeSavingsPlans :: DescribeSavingsPlans

Create a value of DescribeSavingsPlans with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:states:DescribeSavingsPlans', describeSavingsPlans_states - The states.

$sel:savingsPlanIds:DescribeSavingsPlans', describeSavingsPlans_savingsPlanIds - The IDs of the Savings Plans.

$sel:filters:DescribeSavingsPlans', describeSavingsPlans_filters - The filters.

$sel:nextToken:DescribeSavingsPlans', describeSavingsPlans_nextToken - The token for the next page of results.

$sel:savingsPlanArns:DescribeSavingsPlans', describeSavingsPlans_savingsPlanArns - The Amazon Resource Names (ARN) of the Savings Plans.

$sel:maxResults:DescribeSavingsPlans', describeSavingsPlans_maxResults - The maximum number of results to return with a single call. To retrieve additional results, make another call with the returned token value.

data DescribeSavingsPlansResponse

See: newDescribeSavingsPlansResponse smart constructor.

Constructors

DescribeSavingsPlansResponse' (Maybe [SavingsPlan]) (Maybe Text) Int

newDescribeSavingsPlansResponse

Arguments

:: Int
$sel:httpStatus:DescribeSavingsPlansResponse'

-> DescribeSavingsPlansResponse

Create a value of DescribeSavingsPlansResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:savingsPlans:DescribeSavingsPlansResponse', describeSavingsPlansResponse_savingsPlans - Information about the Savings Plans.

$sel:nextToken:DescribeSavingsPlans', describeSavingsPlansResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeSavingsPlansResponse', describeSavingsPlansResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The Amazon Resource Name (ARN) of the resource.

$sel:tags:TagResource', tagResource_tags - One or more tags. For example, { "tags": {"key1":"value1", "key2":"value2"} }.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text [Text]

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - The Amazon Resource Name (ARN) of the resource.

$sel:tagKeys:UntagResource', untagResource_tagKeys - The tag keys.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

DescribeSavingsPlansOfferingRates

data DescribeSavingsPlansOfferingRates

See: newDescribeSavingsPlansOfferingRates smart constructor.

Constructors

DescribeSavingsPlansOfferingRates' (Maybe [Text]) (Maybe [SavingsPlanRateServiceCode]) (Maybe [SavingsPlanOfferingRateFilterElement]) (Maybe Text) (Maybe [SavingsPlanType]) (Maybe [SavingsPlanProductType]) (Maybe [SavingsPlanPaymentOption]) (Maybe [Text]) (Maybe [Text]) (Maybe Natural)

newDescribeSavingsPlansOfferingRates :: DescribeSavingsPlansOfferingRates

Create a value of DescribeSavingsPlansOfferingRates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:savingsPlanOfferingIds:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_savingsPlanOfferingIds - The IDs of the offerings.

$sel:serviceCodes:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_serviceCodes - The services.

$sel:filters:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_filters - The filters.

$sel:nextToken:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_nextToken - The token for the next page of results.

$sel:savingsPlanTypes:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_savingsPlanTypes - The plan types.

$sel:products:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_products - The AWS products.

$sel:savingsPlanPaymentOptions:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_savingsPlanPaymentOptions - The payment options.

$sel:usageTypes:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_usageTypes - The usage details of the line item in the billing report.

$sel:operations:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_operations - The specific AWS operation for the line item in the billing report.

$sel:maxResults:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_maxResults - The maximum number of results to return with a single call. To retrieve additional results, make another call with the returned token value.

data DescribeSavingsPlansOfferingRatesResponse

See: newDescribeSavingsPlansOfferingRatesResponse smart constructor.

Constructors

DescribeSavingsPlansOfferingRatesResponse' (Maybe [SavingsPlanOfferingRate]) (Maybe Text) Int

newDescribeSavingsPlansOfferingRatesResponse

Arguments

:: Int
$sel:httpStatus:DescribeSavingsPlansOfferingRatesResponse'

-> DescribeSavingsPlansOfferingRatesResponse

Create a value of DescribeSavingsPlansOfferingRatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:searchResults:DescribeSavingsPlansOfferingRatesResponse', describeSavingsPlansOfferingRatesResponse_searchResults - Information about the Savings Plans offering rates.

$sel:nextToken:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRatesResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeSavingsPlansOfferingRatesResponse', describeSavingsPlansOfferingRatesResponse_httpStatus - The response's http status code.

Types

CurrencyCode

newtype CurrencyCode

Constructors

CurrencyCode'

Fields

fromCurrencyCode :: Text

Bundled Patterns

pattern CurrencyCode_CNY :: CurrencyCode

pattern CurrencyCode_USD :: CurrencyCode

SavingsPlanOfferingFilterAttribute

newtype SavingsPlanOfferingFilterAttribute

Constructors

SavingsPlanOfferingFilterAttribute'

Fields

fromSavingsPlanOfferingFilterAttribute :: Text

Bundled Patterns

pattern SavingsPlanOfferingFilterAttribute_InstanceFamily :: SavingsPlanOfferingFilterAttribute

pattern SavingsPlanOfferingFilterAttribute_Region :: SavingsPlanOfferingFilterAttribute

SavingsPlanOfferingPropertyKey

newtype SavingsPlanOfferingPropertyKey

Constructors

SavingsPlanOfferingPropertyKey'

Fields

fromSavingsPlanOfferingPropertyKey :: Text

Bundled Patterns

pattern SavingsPlanOfferingPropertyKey_InstanceFamily :: SavingsPlanOfferingPropertyKey

pattern SavingsPlanOfferingPropertyKey_Region :: SavingsPlanOfferingPropertyKey

SavingsPlanPaymentOption

newtype SavingsPlanPaymentOption

Constructors

SavingsPlanPaymentOption'

Fields

fromSavingsPlanPaymentOption :: Text

Bundled Patterns

pattern SavingsPlanPaymentOption_All_Upfront :: SavingsPlanPaymentOption

pattern SavingsPlanPaymentOption_No_Upfront :: SavingsPlanPaymentOption

pattern SavingsPlanPaymentOption_Partial_Upfront :: SavingsPlanPaymentOption

SavingsPlanProductType

newtype SavingsPlanProductType

Constructors

SavingsPlanProductType'

Fields

fromSavingsPlanProductType :: Text

Bundled Patterns

pattern SavingsPlanProductType_EC2 :: SavingsPlanProductType

pattern SavingsPlanProductType_Fargate :: SavingsPlanProductType

pattern SavingsPlanProductType_Lambda :: SavingsPlanProductType

pattern SavingsPlanProductType_SageMaker :: SavingsPlanProductType

SavingsPlanRateFilterAttribute

newtype SavingsPlanRateFilterAttribute

Constructors

SavingsPlanRateFilterAttribute'

Fields

fromSavingsPlanRateFilterAttribute :: Text

Bundled Patterns

pattern SavingsPlanRateFilterAttribute_InstanceFamily :: SavingsPlanRateFilterAttribute

pattern SavingsPlanRateFilterAttribute_InstanceType :: SavingsPlanRateFilterAttribute

pattern SavingsPlanRateFilterAttribute_ProductDescription :: SavingsPlanRateFilterAttribute

pattern SavingsPlanRateFilterAttribute_ProductId :: SavingsPlanRateFilterAttribute

pattern SavingsPlanRateFilterAttribute_Region :: SavingsPlanRateFilterAttribute

pattern SavingsPlanRateFilterAttribute_Tenancy :: SavingsPlanRateFilterAttribute

SavingsPlanRateFilterName

newtype SavingsPlanRateFilterName

Constructors

SavingsPlanRateFilterName'

Fields

fromSavingsPlanRateFilterName :: Text

Bundled Patterns

pattern SavingsPlanRateFilterName_InstanceType :: SavingsPlanRateFilterName

pattern SavingsPlanRateFilterName_Operation :: SavingsPlanRateFilterName

pattern SavingsPlanRateFilterName_ProductDescription :: SavingsPlanRateFilterName

pattern SavingsPlanRateFilterName_ProductType :: SavingsPlanRateFilterName

pattern SavingsPlanRateFilterName_Region :: SavingsPlanRateFilterName

pattern SavingsPlanRateFilterName_ServiceCode :: SavingsPlanRateFilterName

pattern SavingsPlanRateFilterName_Tenancy :: SavingsPlanRateFilterName

pattern SavingsPlanRateFilterName_UsageType :: SavingsPlanRateFilterName

SavingsPlanRatePropertyKey

newtype SavingsPlanRatePropertyKey

Constructors

SavingsPlanRatePropertyKey'

Fields

fromSavingsPlanRatePropertyKey :: Text

Bundled Patterns

pattern SavingsPlanRatePropertyKey_InstanceFamily :: SavingsPlanRatePropertyKey

pattern SavingsPlanRatePropertyKey_InstanceType :: SavingsPlanRatePropertyKey

pattern SavingsPlanRatePropertyKey_ProductDescription :: SavingsPlanRatePropertyKey

pattern SavingsPlanRatePropertyKey_Region :: SavingsPlanRatePropertyKey

pattern SavingsPlanRatePropertyKey_Tenancy :: SavingsPlanRatePropertyKey

SavingsPlanRateServiceCode

newtype SavingsPlanRateServiceCode

Constructors

SavingsPlanRateServiceCode'

Fields

fromSavingsPlanRateServiceCode :: Text

Bundled Patterns

pattern SavingsPlanRateServiceCode_AWSLambda :: SavingsPlanRateServiceCode

pattern SavingsPlanRateServiceCode_AmazonEC2 :: SavingsPlanRateServiceCode

pattern SavingsPlanRateServiceCode_AmazonECS :: SavingsPlanRateServiceCode

pattern SavingsPlanRateServiceCode_AmazonEKS :: SavingsPlanRateServiceCode

pattern SavingsPlanRateServiceCode_AmazonSageMaker :: SavingsPlanRateServiceCode

SavingsPlanRateUnit

newtype SavingsPlanRateUnit

Constructors

SavingsPlanRateUnit'

Fields

fromSavingsPlanRateUnit :: Text

Bundled Patterns

pattern SavingsPlanRateUnit_Hrs :: SavingsPlanRateUnit

pattern SavingsPlanRateUnit_Lambda_GB_Second :: SavingsPlanRateUnit

pattern SavingsPlanRateUnit_Request :: SavingsPlanRateUnit

SavingsPlanState

newtype SavingsPlanState

Constructors

SavingsPlanState'

Fields

fromSavingsPlanState :: Text

Bundled Patterns

pattern SavingsPlanState_Active :: SavingsPlanState

pattern SavingsPlanState_Payment_failed :: SavingsPlanState

pattern SavingsPlanState_Payment_pending :: SavingsPlanState

pattern SavingsPlanState_Queued :: SavingsPlanState

pattern SavingsPlanState_Queued_deleted :: SavingsPlanState

pattern SavingsPlanState_Retired :: SavingsPlanState

SavingsPlanType

newtype SavingsPlanType

Constructors

SavingsPlanType'

Fields

fromSavingsPlanType :: Text

Bundled Patterns

pattern SavingsPlanType_Compute :: SavingsPlanType

pattern SavingsPlanType_EC2Instance :: SavingsPlanType

pattern SavingsPlanType_SageMaker :: SavingsPlanType

SavingsPlansFilterName

newtype SavingsPlansFilterName

Constructors

SavingsPlansFilterName'

Fields

fromSavingsPlansFilterName :: Text

Bundled Patterns

pattern SavingsPlansFilterName_Commitment :: SavingsPlansFilterName

pattern SavingsPlansFilterName_Ec2_instance_family :: SavingsPlansFilterName

pattern SavingsPlansFilterName_End :: SavingsPlansFilterName

pattern SavingsPlansFilterName_Payment_option :: SavingsPlansFilterName

pattern SavingsPlansFilterName_Region :: SavingsPlansFilterName

pattern SavingsPlansFilterName_Savings_plan_type :: SavingsPlansFilterName

pattern SavingsPlansFilterName_Start :: SavingsPlansFilterName

pattern SavingsPlansFilterName_Term :: SavingsPlansFilterName

pattern SavingsPlansFilterName_Upfront :: SavingsPlansFilterName

ParentSavingsPlanOffering

data ParentSavingsPlanOffering

Information about a Savings Plan offering.

See: newParentSavingsPlanOffering smart constructor.

Constructors

ParentSavingsPlanOffering' (Maybe CurrencyCode) (Maybe Natural) (Maybe Text) (Maybe Text) (Maybe SavingsPlanType) (Maybe SavingsPlanPaymentOption)

newParentSavingsPlanOffering :: ParentSavingsPlanOffering

Create a value of ParentSavingsPlanOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currency:ParentSavingsPlanOffering', parentSavingsPlanOffering_currency - The currency.

$sel:durationSeconds:ParentSavingsPlanOffering', parentSavingsPlanOffering_durationSeconds - The duration, in seconds.

$sel:offeringId:ParentSavingsPlanOffering', parentSavingsPlanOffering_offeringId - The ID of the offering.

$sel:planDescription:ParentSavingsPlanOffering', parentSavingsPlanOffering_planDescription - The description.

$sel:planType:ParentSavingsPlanOffering', parentSavingsPlanOffering_planType - The plan type.

$sel:paymentOption:ParentSavingsPlanOffering', parentSavingsPlanOffering_paymentOption - The payment option.

SavingsPlan

data SavingsPlan

Information about a Savings Plan.

See: newSavingsPlan smart constructor.

Constructors

SavingsPlan' (Maybe Integer) (Maybe SavingsPlanType) (Maybe SavingsPlanState) (Maybe [SavingsPlanProductType]) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe CurrencyCode) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe SavingsPlanPaymentOption) (Maybe Text) (Maybe (HashMap Text Text))

newSavingsPlan :: SavingsPlan

Create a value of SavingsPlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:termDurationInSeconds:SavingsPlan', savingsPlan_termDurationInSeconds - The duration of the term, in seconds.

$sel:savingsPlanType:SavingsPlan', savingsPlan_savingsPlanType - The plan type.

$sel:state:SavingsPlan', savingsPlan_state - The state.

$sel:productTypes:SavingsPlan', savingsPlan_productTypes - The product types.

$sel:start:SavingsPlan', savingsPlan_start - The start time.

$sel:ec2InstanceFamily:SavingsPlan', savingsPlan_ec2InstanceFamily - The EC2 instance family.

$sel:savingsPlanId:SavingsPlan', savingsPlan_savingsPlanId - The ID of the Savings Plan.

$sel:recurringPaymentAmount:SavingsPlan', savingsPlan_recurringPaymentAmount - The recurring payment amount.

$sel:currency:SavingsPlan', savingsPlan_currency - The currency.

$sel:end:SavingsPlan', savingsPlan_end - The end time.

$sel:upfrontPaymentAmount:SavingsPlan', savingsPlan_upfrontPaymentAmount - The up-front payment amount.

$sel:offeringId:SavingsPlan', savingsPlan_offeringId - The ID of the offering.

$sel:region:SavingsPlan', savingsPlan_region - The AWS Region.

$sel:commitment:SavingsPlan', savingsPlan_commitment - The hourly commitment, in USD.

$sel:description:SavingsPlan', savingsPlan_description - The description.

$sel:paymentOption:SavingsPlan', savingsPlan_paymentOption - The payment option.

$sel:savingsPlanArn:SavingsPlan', savingsPlan_savingsPlanArn - The Amazon Resource Name (ARN) of the Savings Plan.

$sel:tags:SavingsPlan', savingsPlan_tags - One or more tags.

SavingsPlanFilter

data SavingsPlanFilter

Information about a filter.

See: newSavingsPlanFilter smart constructor.

Constructors

SavingsPlanFilter' (Maybe [Text]) (Maybe SavingsPlansFilterName)

newSavingsPlanFilter :: SavingsPlanFilter

Create a value of SavingsPlanFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:SavingsPlanFilter', savingsPlanFilter_values - The filter value.

$sel:name:SavingsPlanFilter', savingsPlanFilter_name - The filter name.

SavingsPlanOffering

data SavingsPlanOffering

Information about a Savings Plan offering.

See: newSavingsPlanOffering smart constructor.

Constructors

SavingsPlanOffering' (Maybe Text) (Maybe Text) (Maybe [SavingsPlanProductType]) (Maybe CurrencyCode) (Maybe Natural) (Maybe Text) (Maybe Text) (Maybe SavingsPlanType) (Maybe Text) (Maybe SavingsPlanPaymentOption) (Maybe [SavingsPlanOfferingProperty])

newSavingsPlanOffering :: SavingsPlanOffering

Create a value of SavingsPlanOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operation:SavingsPlanOffering', savingsPlanOffering_operation - The specific AWS operation for the line item in the billing report.

$sel:usageType:SavingsPlanOffering', savingsPlanOffering_usageType - The usage details of the line item in the billing report.

$sel:productTypes:SavingsPlanOffering', savingsPlanOffering_productTypes - The product type.

$sel:currency:SavingsPlanOffering', savingsPlanOffering_currency - The currency.

$sel:durationSeconds:SavingsPlanOffering', savingsPlanOffering_durationSeconds - The duration, in seconds.

$sel:offeringId:SavingsPlanOffering', savingsPlanOffering_offeringId - The ID of the offering.

$sel:serviceCode:SavingsPlanOffering', savingsPlanOffering_serviceCode - The service.

$sel:planType:SavingsPlanOffering', savingsPlanOffering_planType - The plan type.

$sel:description:SavingsPlanOffering', savingsPlanOffering_description - The description.

$sel:paymentOption:SavingsPlanOffering', savingsPlanOffering_paymentOption - The payment option.

$sel:properties:SavingsPlanOffering', savingsPlanOffering_properties - The properties.

SavingsPlanOfferingFilterElement

data SavingsPlanOfferingFilterElement

Information about a filter.

See: newSavingsPlanOfferingFilterElement smart constructor.

Constructors

SavingsPlanOfferingFilterElement' (Maybe [Text]) (Maybe SavingsPlanOfferingFilterAttribute)

newSavingsPlanOfferingFilterElement :: SavingsPlanOfferingFilterElement

Create a value of SavingsPlanOfferingFilterElement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:SavingsPlanOfferingFilterElement', savingsPlanOfferingFilterElement_values - The filter values.

$sel:name:SavingsPlanOfferingFilterElement', savingsPlanOfferingFilterElement_name - The filter name.

SavingsPlanOfferingProperty

data SavingsPlanOfferingProperty

Information about a property.

See: newSavingsPlanOfferingProperty smart constructor.

Constructors

SavingsPlanOfferingProperty' (Maybe Text) (Maybe SavingsPlanOfferingPropertyKey)

newSavingsPlanOfferingProperty :: SavingsPlanOfferingProperty

Create a value of SavingsPlanOfferingProperty with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:SavingsPlanOfferingProperty', savingsPlanOfferingProperty_value - The property value.

$sel:name:SavingsPlanOfferingProperty', savingsPlanOfferingProperty_name - The property name.

SavingsPlanOfferingRate

data SavingsPlanOfferingRate

Information about a Savings Plan offering rate.

See: newSavingsPlanOfferingRate smart constructor.

Constructors

SavingsPlanOfferingRate' (Maybe Text) (Maybe Text) (Maybe SavingsPlanProductType) (Maybe ParentSavingsPlanOffering) (Maybe Text) (Maybe SavingsPlanRateServiceCode) (Maybe SavingsPlanRateUnit) (Maybe [SavingsPlanOfferingRateProperty])

newSavingsPlanOfferingRate :: SavingsPlanOfferingRate

Create a value of SavingsPlanOfferingRate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operation:SavingsPlanOfferingRate', savingsPlanOfferingRate_operation - The specific AWS operation for the line item in the billing report.

$sel:usageType:SavingsPlanOfferingRate', savingsPlanOfferingRate_usageType - The usage details of the line item in the billing report.

$sel:productType:SavingsPlanOfferingRate', savingsPlanOfferingRate_productType - The product type.

$sel:savingsPlanOffering:SavingsPlanOfferingRate', savingsPlanOfferingRate_savingsPlanOffering - The Savings Plan offering.

$sel:rate:SavingsPlanOfferingRate', savingsPlanOfferingRate_rate - The Savings Plan rate.

$sel:serviceCode:SavingsPlanOfferingRate', savingsPlanOfferingRate_serviceCode - The service.

$sel:unit:SavingsPlanOfferingRate', savingsPlanOfferingRate_unit - The unit.

$sel:properties:SavingsPlanOfferingRate', savingsPlanOfferingRate_properties - The properties.

SavingsPlanOfferingRateFilterElement

data SavingsPlanOfferingRateFilterElement

Information about a filter.

See: newSavingsPlanOfferingRateFilterElement smart constructor.

Constructors

SavingsPlanOfferingRateFilterElement' (Maybe [Text]) (Maybe SavingsPlanRateFilterAttribute)

newSavingsPlanOfferingRateFilterElement :: SavingsPlanOfferingRateFilterElement

Create a value of SavingsPlanOfferingRateFilterElement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:SavingsPlanOfferingRateFilterElement', savingsPlanOfferingRateFilterElement_values - The filter values.

$sel:name:SavingsPlanOfferingRateFilterElement', savingsPlanOfferingRateFilterElement_name - The filter name.

SavingsPlanOfferingRateProperty

data SavingsPlanOfferingRateProperty

Information about a property.

See: newSavingsPlanOfferingRateProperty smart constructor.

Constructors

SavingsPlanOfferingRateProperty' (Maybe Text) (Maybe Text)

newSavingsPlanOfferingRateProperty :: SavingsPlanOfferingRateProperty

Create a value of SavingsPlanOfferingRateProperty with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:SavingsPlanOfferingRateProperty', savingsPlanOfferingRateProperty_value - The property value.

$sel:name:SavingsPlanOfferingRateProperty', savingsPlanOfferingRateProperty_name - The property name.

SavingsPlanRate

data SavingsPlanRate

Information about a Savings Plan rate.

See: newSavingsPlanRate smart constructor.

Constructors

SavingsPlanRate' (Maybe Text) (Maybe Text) (Maybe SavingsPlanProductType) (Maybe CurrencyCode) (Maybe Text) (Maybe SavingsPlanRateServiceCode) (Maybe SavingsPlanRateUnit) (Maybe [SavingsPlanRateProperty])

newSavingsPlanRate :: SavingsPlanRate

Create a value of SavingsPlanRate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operation:SavingsPlanRate', savingsPlanRate_operation - The specific AWS operation for the line item in the billing report.

$sel:usageType:SavingsPlanRate', savingsPlanRate_usageType - The usage details of the line item in the billing report.

$sel:productType:SavingsPlanRate', savingsPlanRate_productType - The product type.

$sel:currency:SavingsPlanRate', savingsPlanRate_currency - The currency.

$sel:rate:SavingsPlanRate', savingsPlanRate_rate - The rate.

$sel:serviceCode:SavingsPlanRate', savingsPlanRate_serviceCode - The service.

$sel:unit:SavingsPlanRate', savingsPlanRate_unit - The unit.

$sel:properties:SavingsPlanRate', savingsPlanRate_properties - The properties.

SavingsPlanRateFilter

data SavingsPlanRateFilter

Information about a filter.

See: newSavingsPlanRateFilter smart constructor.

Constructors

SavingsPlanRateFilter' (Maybe [Text]) (Maybe SavingsPlanRateFilterName)

newSavingsPlanRateFilter :: SavingsPlanRateFilter

Create a value of SavingsPlanRateFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:SavingsPlanRateFilter', savingsPlanRateFilter_values - The filter values.

$sel:name:SavingsPlanRateFilter', savingsPlanRateFilter_name - The filter name.

SavingsPlanRateProperty

data SavingsPlanRateProperty

Information about a property.

See: newSavingsPlanRateProperty smart constructor.

Constructors

SavingsPlanRateProperty' (Maybe Text) (Maybe SavingsPlanRatePropertyKey)

newSavingsPlanRateProperty :: SavingsPlanRateProperty

Create a value of SavingsPlanRateProperty with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:SavingsPlanRateProperty', savingsPlanRateProperty_value - The property value.

$sel:name:SavingsPlanRateProperty', savingsPlanRateProperty_name - The property name.