Network.AWS.SavingsPlans.DescribeSavingsPlansOfferingRates

Description

Describes the specified Savings Plans offering rates.

Creating a Request

data DescribeSavingsPlansOfferingRates

See: newDescribeSavingsPlansOfferingRates smart constructor.

Constructors

DescribeSavingsPlansOfferingRates'

Fields

savingsPlanOfferingIds :: Maybe [Text]

The IDs of the offerings.

serviceCodes :: Maybe [SavingsPlanRateServiceCode]

The services.

filters :: Maybe [SavingsPlanOfferingRateFilterElement]

The filters.

nextToken :: Maybe Text

The token for the next page of results.

savingsPlanTypes :: Maybe [SavingsPlanType]

The plan types.

products :: Maybe [SavingsPlanProductType]

The AWS products.

savingsPlanPaymentOptions :: Maybe [SavingsPlanPaymentOption]

The payment options.

usageTypes :: Maybe [Text]

The usage details of the line item in the billing report.

operations :: Maybe [Text]

The specific AWS operation for the line item in the billing report.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve additional results, make another call with the returned token value.

newDescribeSavingsPlansOfferingRates :: DescribeSavingsPlansOfferingRates

Create a value of DescribeSavingsPlansOfferingRates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:savingsPlanOfferingIds:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_savingsPlanOfferingIds - The IDs of the offerings.

$sel:serviceCodes:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_serviceCodes - The services.

$sel:filters:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_filters - The filters.

$sel:nextToken:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_nextToken - The token for the next page of results.

$sel:savingsPlanTypes:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_savingsPlanTypes - The plan types.

$sel:products:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_products - The AWS products.

$sel:savingsPlanPaymentOptions:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_savingsPlanPaymentOptions - The payment options.

$sel:usageTypes:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_usageTypes - The usage details of the line item in the billing report.

$sel:operations:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_operations - The specific AWS operation for the line item in the billing report.

$sel:maxResults:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRates_maxResults - The maximum number of results to return with a single call. To retrieve additional results, make another call with the returned token value.

Request Lenses

describeSavingsPlansOfferingRates_savingsPlanOfferingIds :: Lens' DescribeSavingsPlansOfferingRates (Maybe [Text])

The IDs of the offerings.

describeSavingsPlansOfferingRates_serviceCodes :: Lens' DescribeSavingsPlansOfferingRates (Maybe [SavingsPlanRateServiceCode])

The services.

describeSavingsPlansOfferingRates_filters :: Lens' DescribeSavingsPlansOfferingRates (Maybe [SavingsPlanOfferingRateFilterElement])

The filters.

describeSavingsPlansOfferingRates_nextToken :: Lens' DescribeSavingsPlansOfferingRates (Maybe Text)

The token for the next page of results.

describeSavingsPlansOfferingRates_savingsPlanTypes :: Lens' DescribeSavingsPlansOfferingRates (Maybe [SavingsPlanType])

The plan types.

describeSavingsPlansOfferingRates_products :: Lens' DescribeSavingsPlansOfferingRates (Maybe [SavingsPlanProductType])

The AWS products.

describeSavingsPlansOfferingRates_savingsPlanPaymentOptions :: Lens' DescribeSavingsPlansOfferingRates (Maybe [SavingsPlanPaymentOption])

The payment options.

describeSavingsPlansOfferingRates_usageTypes :: Lens' DescribeSavingsPlansOfferingRates (Maybe [Text])

The usage details of the line item in the billing report.

describeSavingsPlansOfferingRates_operations :: Lens' DescribeSavingsPlansOfferingRates (Maybe [Text])

The specific AWS operation for the line item in the billing report.

describeSavingsPlansOfferingRates_maxResults :: Lens' DescribeSavingsPlansOfferingRates (Maybe Natural)

The maximum number of results to return with a single call. To retrieve additional results, make another call with the returned token value.

Destructuring the Response

data DescribeSavingsPlansOfferingRatesResponse

See: newDescribeSavingsPlansOfferingRatesResponse smart constructor.

Constructors

DescribeSavingsPlansOfferingRatesResponse'

Fields

searchResults :: Maybe [SavingsPlanOfferingRate]

Information about the Savings Plans offering rates.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeSavingsPlansOfferingRatesResponse

Arguments

:: Int
$sel:httpStatus:DescribeSavingsPlansOfferingRatesResponse'

-> DescribeSavingsPlansOfferingRatesResponse

Create a value of DescribeSavingsPlansOfferingRatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:searchResults:DescribeSavingsPlansOfferingRatesResponse', describeSavingsPlansOfferingRatesResponse_searchResults - Information about the Savings Plans offering rates.

$sel:nextToken:DescribeSavingsPlansOfferingRates', describeSavingsPlansOfferingRatesResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeSavingsPlansOfferingRatesResponse', describeSavingsPlansOfferingRatesResponse_httpStatus - The response's http status code.

Response Lenses

describeSavingsPlansOfferingRatesResponse_searchResults :: Lens' DescribeSavingsPlansOfferingRatesResponse (Maybe [SavingsPlanOfferingRate])

Information about the Savings Plans offering rates.

describeSavingsPlansOfferingRatesResponse_nextToken :: Lens' DescribeSavingsPlansOfferingRatesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeSavingsPlansOfferingRatesResponse_httpStatus :: Lens' DescribeSavingsPlansOfferingRatesResponse Int

The response's http status code.