Network.AWS.SavingsPlans.Types

Description

Service Configuration

defaultService :: Service

API version 2019-06-28 of the Amazon Savings Plans SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

One of the input parameters is not valid.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A service quota has been exceeded.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

An unexpected error occurred.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource was not found.

CurrencyCode

newtype CurrencyCode

Constructors

CurrencyCode'

Fields

fromCurrencyCode :: Text

Bundled Patterns

pattern CurrencyCode_CNY :: CurrencyCode

pattern CurrencyCode_USD :: CurrencyCode

SavingsPlanOfferingFilterAttribute

newtype SavingsPlanOfferingFilterAttribute

Constructors

SavingsPlanOfferingFilterAttribute'

Fields

fromSavingsPlanOfferingFilterAttribute :: Text

Bundled Patterns

pattern SavingsPlanOfferingFilterAttribute_InstanceFamily :: SavingsPlanOfferingFilterAttribute

pattern SavingsPlanOfferingFilterAttribute_Region :: SavingsPlanOfferingFilterAttribute

SavingsPlanOfferingPropertyKey

newtype SavingsPlanOfferingPropertyKey

Constructors

SavingsPlanOfferingPropertyKey'

Fields

fromSavingsPlanOfferingPropertyKey :: Text

Bundled Patterns

pattern SavingsPlanOfferingPropertyKey_InstanceFamily :: SavingsPlanOfferingPropertyKey

pattern SavingsPlanOfferingPropertyKey_Region :: SavingsPlanOfferingPropertyKey

SavingsPlanPaymentOption

newtype SavingsPlanPaymentOption

Constructors

SavingsPlanPaymentOption'

Fields

fromSavingsPlanPaymentOption :: Text

Bundled Patterns

pattern SavingsPlanPaymentOption_All_Upfront :: SavingsPlanPaymentOption

pattern SavingsPlanPaymentOption_No_Upfront :: SavingsPlanPaymentOption

pattern SavingsPlanPaymentOption_Partial_Upfront :: SavingsPlanPaymentOption

SavingsPlanProductType

newtype SavingsPlanProductType

Constructors

SavingsPlanProductType'

Fields

fromSavingsPlanProductType :: Text

Bundled Patterns

pattern SavingsPlanProductType_EC2 :: SavingsPlanProductType

pattern SavingsPlanProductType_Fargate :: SavingsPlanProductType

pattern SavingsPlanProductType_Lambda :: SavingsPlanProductType

pattern SavingsPlanProductType_SageMaker :: SavingsPlanProductType

SavingsPlanRateFilterAttribute

newtype SavingsPlanRateFilterAttribute

Constructors

SavingsPlanRateFilterAttribute'

Fields

fromSavingsPlanRateFilterAttribute :: Text

Bundled Patterns

pattern SavingsPlanRateFilterAttribute_InstanceFamily :: SavingsPlanRateFilterAttribute

pattern SavingsPlanRateFilterAttribute_InstanceType :: SavingsPlanRateFilterAttribute

pattern SavingsPlanRateFilterAttribute_ProductDescription :: SavingsPlanRateFilterAttribute

pattern SavingsPlanRateFilterAttribute_ProductId :: SavingsPlanRateFilterAttribute

pattern SavingsPlanRateFilterAttribute_Region :: SavingsPlanRateFilterAttribute

pattern SavingsPlanRateFilterAttribute_Tenancy :: SavingsPlanRateFilterAttribute

SavingsPlanRateFilterName

newtype SavingsPlanRateFilterName

Constructors

SavingsPlanRateFilterName'

Fields

fromSavingsPlanRateFilterName :: Text

Bundled Patterns

pattern SavingsPlanRateFilterName_InstanceType :: SavingsPlanRateFilterName

pattern SavingsPlanRateFilterName_Operation :: SavingsPlanRateFilterName

pattern SavingsPlanRateFilterName_ProductDescription :: SavingsPlanRateFilterName

pattern SavingsPlanRateFilterName_ProductType :: SavingsPlanRateFilterName

pattern SavingsPlanRateFilterName_Region :: SavingsPlanRateFilterName

pattern SavingsPlanRateFilterName_ServiceCode :: SavingsPlanRateFilterName

pattern SavingsPlanRateFilterName_Tenancy :: SavingsPlanRateFilterName

pattern SavingsPlanRateFilterName_UsageType :: SavingsPlanRateFilterName

SavingsPlanRatePropertyKey

newtype SavingsPlanRatePropertyKey

Constructors

SavingsPlanRatePropertyKey'

Fields

fromSavingsPlanRatePropertyKey :: Text

Bundled Patterns

pattern SavingsPlanRatePropertyKey_InstanceFamily :: SavingsPlanRatePropertyKey

pattern SavingsPlanRatePropertyKey_InstanceType :: SavingsPlanRatePropertyKey

pattern SavingsPlanRatePropertyKey_ProductDescription :: SavingsPlanRatePropertyKey

pattern SavingsPlanRatePropertyKey_Region :: SavingsPlanRatePropertyKey

pattern SavingsPlanRatePropertyKey_Tenancy :: SavingsPlanRatePropertyKey

SavingsPlanRateServiceCode

newtype SavingsPlanRateServiceCode

Constructors

SavingsPlanRateServiceCode'

Fields

fromSavingsPlanRateServiceCode :: Text

Bundled Patterns

pattern SavingsPlanRateServiceCode_AWSLambda :: SavingsPlanRateServiceCode

pattern SavingsPlanRateServiceCode_AmazonEC2 :: SavingsPlanRateServiceCode

pattern SavingsPlanRateServiceCode_AmazonECS :: SavingsPlanRateServiceCode

pattern SavingsPlanRateServiceCode_AmazonEKS :: SavingsPlanRateServiceCode

pattern SavingsPlanRateServiceCode_AmazonSageMaker :: SavingsPlanRateServiceCode

SavingsPlanRateUnit

newtype SavingsPlanRateUnit

Constructors

SavingsPlanRateUnit'

Fields

fromSavingsPlanRateUnit :: Text

Bundled Patterns

pattern SavingsPlanRateUnit_Hrs :: SavingsPlanRateUnit

pattern SavingsPlanRateUnit_Lambda_GB_Second :: SavingsPlanRateUnit

pattern SavingsPlanRateUnit_Request :: SavingsPlanRateUnit

SavingsPlanState

newtype SavingsPlanState

Constructors

SavingsPlanState'

Fields

fromSavingsPlanState :: Text

Bundled Patterns

pattern SavingsPlanState_Active :: SavingsPlanState

pattern SavingsPlanState_Payment_failed :: SavingsPlanState

pattern SavingsPlanState_Payment_pending :: SavingsPlanState

pattern SavingsPlanState_Queued :: SavingsPlanState

pattern SavingsPlanState_Queued_deleted :: SavingsPlanState

pattern SavingsPlanState_Retired :: SavingsPlanState

SavingsPlanType

newtype SavingsPlanType

Constructors

SavingsPlanType'

Fields

fromSavingsPlanType :: Text

Bundled Patterns

pattern SavingsPlanType_Compute :: SavingsPlanType

pattern SavingsPlanType_EC2Instance :: SavingsPlanType

pattern SavingsPlanType_SageMaker :: SavingsPlanType

SavingsPlansFilterName

newtype SavingsPlansFilterName

Constructors

SavingsPlansFilterName'

Fields

fromSavingsPlansFilterName :: Text

Bundled Patterns

pattern SavingsPlansFilterName_Commitment :: SavingsPlansFilterName

pattern SavingsPlansFilterName_Ec2_instance_family :: SavingsPlansFilterName

pattern SavingsPlansFilterName_End :: SavingsPlansFilterName

pattern SavingsPlansFilterName_Payment_option :: SavingsPlansFilterName

pattern SavingsPlansFilterName_Region :: SavingsPlansFilterName

pattern SavingsPlansFilterName_Savings_plan_type :: SavingsPlansFilterName

pattern SavingsPlansFilterName_Start :: SavingsPlansFilterName

pattern SavingsPlansFilterName_Term :: SavingsPlansFilterName

pattern SavingsPlansFilterName_Upfront :: SavingsPlansFilterName

ParentSavingsPlanOffering

data ParentSavingsPlanOffering

Information about a Savings Plan offering.

See: newParentSavingsPlanOffering smart constructor.

Constructors

ParentSavingsPlanOffering'

Fields

currency :: Maybe CurrencyCode

The currency.

durationSeconds :: Maybe Natural

The duration, in seconds.

offeringId :: Maybe Text

The ID of the offering.

planDescription :: Maybe Text

The description.

planType :: Maybe SavingsPlanType

The plan type.

paymentOption :: Maybe SavingsPlanPaymentOption

The payment option.

newParentSavingsPlanOffering :: ParentSavingsPlanOffering

Create a value of ParentSavingsPlanOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currency:ParentSavingsPlanOffering', parentSavingsPlanOffering_currency - The currency.

$sel:durationSeconds:ParentSavingsPlanOffering', parentSavingsPlanOffering_durationSeconds - The duration, in seconds.

$sel:offeringId:ParentSavingsPlanOffering', parentSavingsPlanOffering_offeringId - The ID of the offering.

$sel:planDescription:ParentSavingsPlanOffering', parentSavingsPlanOffering_planDescription - The description.

$sel:planType:ParentSavingsPlanOffering', parentSavingsPlanOffering_planType - The plan type.

$sel:paymentOption:ParentSavingsPlanOffering', parentSavingsPlanOffering_paymentOption - The payment option.

parentSavingsPlanOffering_currency :: Lens' ParentSavingsPlanOffering (Maybe CurrencyCode)

The currency.

parentSavingsPlanOffering_durationSeconds :: Lens' ParentSavingsPlanOffering (Maybe Natural)

The duration, in seconds.

parentSavingsPlanOffering_offeringId :: Lens' ParentSavingsPlanOffering (Maybe Text)

The ID of the offering.

parentSavingsPlanOffering_planDescription :: Lens' ParentSavingsPlanOffering (Maybe Text)

The description.

parentSavingsPlanOffering_planType :: Lens' ParentSavingsPlanOffering (Maybe SavingsPlanType)

The plan type.

parentSavingsPlanOffering_paymentOption :: Lens' ParentSavingsPlanOffering (Maybe SavingsPlanPaymentOption)

The payment option.

SavingsPlan

data SavingsPlan

Information about a Savings Plan.

See: newSavingsPlan smart constructor.

Constructors

SavingsPlan'

Fields

termDurationInSeconds :: Maybe Integer

The duration of the term, in seconds.

savingsPlanType :: Maybe SavingsPlanType

The plan type.

state :: Maybe SavingsPlanState

The state.

productTypes :: Maybe [SavingsPlanProductType]

The product types.

start :: Maybe Text

The start time.

ec2InstanceFamily :: Maybe Text

The EC2 instance family.

savingsPlanId :: Maybe Text

The ID of the Savings Plan.

recurringPaymentAmount :: Maybe Text

The recurring payment amount.

currency :: Maybe CurrencyCode

The currency.

end :: Maybe Text

The end time.

upfrontPaymentAmount :: Maybe Text

The up-front payment amount.

offeringId :: Maybe Text

The ID of the offering.

region :: Maybe Text

The AWS Region.

commitment :: Maybe Text

The hourly commitment, in USD.

description :: Maybe Text

The description.

paymentOption :: Maybe SavingsPlanPaymentOption

The payment option.

savingsPlanArn :: Maybe Text

The Amazon Resource Name (ARN) of the Savings Plan.

tags :: Maybe (HashMap Text Text)

One or more tags.

newSavingsPlan :: SavingsPlan

Create a value of SavingsPlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:termDurationInSeconds:SavingsPlan', savingsPlan_termDurationInSeconds - The duration of the term, in seconds.

$sel:savingsPlanType:SavingsPlan', savingsPlan_savingsPlanType - The plan type.

$sel:state:SavingsPlan', savingsPlan_state - The state.

$sel:productTypes:SavingsPlan', savingsPlan_productTypes - The product types.

$sel:start:SavingsPlan', savingsPlan_start - The start time.

$sel:ec2InstanceFamily:SavingsPlan', savingsPlan_ec2InstanceFamily - The EC2 instance family.

$sel:savingsPlanId:SavingsPlan', savingsPlan_savingsPlanId - The ID of the Savings Plan.

$sel:recurringPaymentAmount:SavingsPlan', savingsPlan_recurringPaymentAmount - The recurring payment amount.

$sel:currency:SavingsPlan', savingsPlan_currency - The currency.

$sel:end:SavingsPlan', savingsPlan_end - The end time.

$sel:upfrontPaymentAmount:SavingsPlan', savingsPlan_upfrontPaymentAmount - The up-front payment amount.

$sel:offeringId:SavingsPlan', savingsPlan_offeringId - The ID of the offering.

$sel:region:SavingsPlan', savingsPlan_region - The AWS Region.

$sel:commitment:SavingsPlan', savingsPlan_commitment - The hourly commitment, in USD.

$sel:description:SavingsPlan', savingsPlan_description - The description.

$sel:paymentOption:SavingsPlan', savingsPlan_paymentOption - The payment option.

$sel:savingsPlanArn:SavingsPlan', savingsPlan_savingsPlanArn - The Amazon Resource Name (ARN) of the Savings Plan.

$sel:tags:SavingsPlan', savingsPlan_tags - One or more tags.

savingsPlan_termDurationInSeconds :: Lens' SavingsPlan (Maybe Integer)

The duration of the term, in seconds.

savingsPlan_savingsPlanType :: Lens' SavingsPlan (Maybe SavingsPlanType)

The plan type.

savingsPlan_state :: Lens' SavingsPlan (Maybe SavingsPlanState)

The state.

savingsPlan_productTypes :: Lens' SavingsPlan (Maybe [SavingsPlanProductType])

The product types.

savingsPlan_start :: Lens' SavingsPlan (Maybe Text)

The start time.

savingsPlan_ec2InstanceFamily :: Lens' SavingsPlan (Maybe Text)

The EC2 instance family.

savingsPlan_savingsPlanId :: Lens' SavingsPlan (Maybe Text)

The ID of the Savings Plan.

savingsPlan_recurringPaymentAmount :: Lens' SavingsPlan (Maybe Text)

The recurring payment amount.

savingsPlan_currency :: Lens' SavingsPlan (Maybe CurrencyCode)

The currency.

savingsPlan_end :: Lens' SavingsPlan (Maybe Text)

The end time.

savingsPlan_upfrontPaymentAmount :: Lens' SavingsPlan (Maybe Text)

The up-front payment amount.

savingsPlan_offeringId :: Lens' SavingsPlan (Maybe Text)

The ID of the offering.

savingsPlan_region :: Lens' SavingsPlan (Maybe Text)

The AWS Region.

savingsPlan_commitment :: Lens' SavingsPlan (Maybe Text)

The hourly commitment, in USD.

savingsPlan_description :: Lens' SavingsPlan (Maybe Text)

The description.

savingsPlan_paymentOption :: Lens' SavingsPlan (Maybe SavingsPlanPaymentOption)

The payment option.

savingsPlan_savingsPlanArn :: Lens' SavingsPlan (Maybe Text)

The Amazon Resource Name (ARN) of the Savings Plan.

savingsPlan_tags :: Lens' SavingsPlan (Maybe (HashMap Text Text))

One or more tags.

SavingsPlanFilter

data SavingsPlanFilter

Information about a filter.

See: newSavingsPlanFilter smart constructor.

Constructors

SavingsPlanFilter'

Fields

values :: Maybe [Text]

The filter value.

name :: Maybe SavingsPlansFilterName

The filter name.

newSavingsPlanFilter :: SavingsPlanFilter

Create a value of SavingsPlanFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:SavingsPlanFilter', savingsPlanFilter_values - The filter value.

$sel:name:SavingsPlanFilter', savingsPlanFilter_name - The filter name.

savingsPlanFilter_values :: Lens' SavingsPlanFilter (Maybe [Text])

The filter value.

savingsPlanFilter_name :: Lens' SavingsPlanFilter (Maybe SavingsPlansFilterName)

The filter name.

SavingsPlanOffering

data SavingsPlanOffering

Information about a Savings Plan offering.

See: newSavingsPlanOffering smart constructor.

Constructors

SavingsPlanOffering'

Fields

operation :: Maybe Text

The specific AWS operation for the line item in the billing report.

usageType :: Maybe Text

The usage details of the line item in the billing report.

productTypes :: Maybe [SavingsPlanProductType]

The product type.

currency :: Maybe CurrencyCode

The currency.

durationSeconds :: Maybe Natural

The duration, in seconds.

offeringId :: Maybe Text

The ID of the offering.

serviceCode :: Maybe Text

The service.

planType :: Maybe SavingsPlanType

The plan type.

description :: Maybe Text

The description.

paymentOption :: Maybe SavingsPlanPaymentOption

The payment option.

properties :: Maybe [SavingsPlanOfferingProperty]

The properties.

newSavingsPlanOffering :: SavingsPlanOffering

Create a value of SavingsPlanOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operation:SavingsPlanOffering', savingsPlanOffering_operation - The specific AWS operation for the line item in the billing report.

$sel:usageType:SavingsPlanOffering', savingsPlanOffering_usageType - The usage details of the line item in the billing report.

$sel:productTypes:SavingsPlanOffering', savingsPlanOffering_productTypes - The product type.

$sel:currency:SavingsPlanOffering', savingsPlanOffering_currency - The currency.

$sel:durationSeconds:SavingsPlanOffering', savingsPlanOffering_durationSeconds - The duration, in seconds.

$sel:offeringId:SavingsPlanOffering', savingsPlanOffering_offeringId - The ID of the offering.

$sel:serviceCode:SavingsPlanOffering', savingsPlanOffering_serviceCode - The service.

$sel:planType:SavingsPlanOffering', savingsPlanOffering_planType - The plan type.

$sel:description:SavingsPlanOffering', savingsPlanOffering_description - The description.

$sel:paymentOption:SavingsPlanOffering', savingsPlanOffering_paymentOption - The payment option.

$sel:properties:SavingsPlanOffering', savingsPlanOffering_properties - The properties.

savingsPlanOffering_operation :: Lens' SavingsPlanOffering (Maybe Text)

The specific AWS operation for the line item in the billing report.

savingsPlanOffering_usageType :: Lens' SavingsPlanOffering (Maybe Text)

The usage details of the line item in the billing report.

savingsPlanOffering_productTypes :: Lens' SavingsPlanOffering (Maybe [SavingsPlanProductType])

The product type.

savingsPlanOffering_currency :: Lens' SavingsPlanOffering (Maybe CurrencyCode)

The currency.

savingsPlanOffering_durationSeconds :: Lens' SavingsPlanOffering (Maybe Natural)

The duration, in seconds.

savingsPlanOffering_offeringId :: Lens' SavingsPlanOffering (Maybe Text)

The ID of the offering.

savingsPlanOffering_serviceCode :: Lens' SavingsPlanOffering (Maybe Text)

The service.

savingsPlanOffering_planType :: Lens' SavingsPlanOffering (Maybe SavingsPlanType)

The plan type.

savingsPlanOffering_description :: Lens' SavingsPlanOffering (Maybe Text)

The description.

savingsPlanOffering_paymentOption :: Lens' SavingsPlanOffering (Maybe SavingsPlanPaymentOption)

The payment option.

savingsPlanOffering_properties :: Lens' SavingsPlanOffering (Maybe [SavingsPlanOfferingProperty])

The properties.

SavingsPlanOfferingFilterElement

data SavingsPlanOfferingFilterElement

Information about a filter.

See: newSavingsPlanOfferingFilterElement smart constructor.

Constructors

SavingsPlanOfferingFilterElement'

Fields

values :: Maybe [Text]

The filter values.

name :: Maybe SavingsPlanOfferingFilterAttribute

The filter name.

newSavingsPlanOfferingFilterElement :: SavingsPlanOfferingFilterElement

Create a value of SavingsPlanOfferingFilterElement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:SavingsPlanOfferingFilterElement', savingsPlanOfferingFilterElement_values - The filter values.

$sel:name:SavingsPlanOfferingFilterElement', savingsPlanOfferingFilterElement_name - The filter name.

savingsPlanOfferingFilterElement_values :: Lens' SavingsPlanOfferingFilterElement (Maybe [Text])

The filter values.

savingsPlanOfferingFilterElement_name :: Lens' SavingsPlanOfferingFilterElement (Maybe SavingsPlanOfferingFilterAttribute)

The filter name.

SavingsPlanOfferingProperty

data SavingsPlanOfferingProperty

Information about a property.

See: newSavingsPlanOfferingProperty smart constructor.

Constructors

SavingsPlanOfferingProperty'

Fields

value :: Maybe Text

The property value.

name :: Maybe SavingsPlanOfferingPropertyKey

The property name.

newSavingsPlanOfferingProperty :: SavingsPlanOfferingProperty

Create a value of SavingsPlanOfferingProperty with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:SavingsPlanOfferingProperty', savingsPlanOfferingProperty_value - The property value.

$sel:name:SavingsPlanOfferingProperty', savingsPlanOfferingProperty_name - The property name.

savingsPlanOfferingProperty_value :: Lens' SavingsPlanOfferingProperty (Maybe Text)

The property value.

savingsPlanOfferingProperty_name :: Lens' SavingsPlanOfferingProperty (Maybe SavingsPlanOfferingPropertyKey)

The property name.

SavingsPlanOfferingRate

data SavingsPlanOfferingRate

Information about a Savings Plan offering rate.

See: newSavingsPlanOfferingRate smart constructor.

Constructors

SavingsPlanOfferingRate'

Fields

operation :: Maybe Text

The specific AWS operation for the line item in the billing report.

usageType :: Maybe Text

The usage details of the line item in the billing report.

productType :: Maybe SavingsPlanProductType

The product type.

savingsPlanOffering :: Maybe ParentSavingsPlanOffering

The Savings Plan offering.

rate :: Maybe Text

The Savings Plan rate.

serviceCode :: Maybe SavingsPlanRateServiceCode

The service.

unit :: Maybe SavingsPlanRateUnit

The unit.

properties :: Maybe [SavingsPlanOfferingRateProperty]

The properties.

newSavingsPlanOfferingRate :: SavingsPlanOfferingRate

Create a value of SavingsPlanOfferingRate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operation:SavingsPlanOfferingRate', savingsPlanOfferingRate_operation - The specific AWS operation for the line item in the billing report.

$sel:usageType:SavingsPlanOfferingRate', savingsPlanOfferingRate_usageType - The usage details of the line item in the billing report.

$sel:productType:SavingsPlanOfferingRate', savingsPlanOfferingRate_productType - The product type.

$sel:savingsPlanOffering:SavingsPlanOfferingRate', savingsPlanOfferingRate_savingsPlanOffering - The Savings Plan offering.

$sel:rate:SavingsPlanOfferingRate', savingsPlanOfferingRate_rate - The Savings Plan rate.

$sel:serviceCode:SavingsPlanOfferingRate', savingsPlanOfferingRate_serviceCode - The service.

$sel:unit:SavingsPlanOfferingRate', savingsPlanOfferingRate_unit - The unit.

$sel:properties:SavingsPlanOfferingRate', savingsPlanOfferingRate_properties - The properties.

savingsPlanOfferingRate_operation :: Lens' SavingsPlanOfferingRate (Maybe Text)

The specific AWS operation for the line item in the billing report.

savingsPlanOfferingRate_usageType :: Lens' SavingsPlanOfferingRate (Maybe Text)

The usage details of the line item in the billing report.

savingsPlanOfferingRate_productType :: Lens' SavingsPlanOfferingRate (Maybe SavingsPlanProductType)

The product type.

savingsPlanOfferingRate_savingsPlanOffering :: Lens' SavingsPlanOfferingRate (Maybe ParentSavingsPlanOffering)

The Savings Plan offering.

savingsPlanOfferingRate_rate :: Lens' SavingsPlanOfferingRate (Maybe Text)

The Savings Plan rate.

savingsPlanOfferingRate_serviceCode :: Lens' SavingsPlanOfferingRate (Maybe SavingsPlanRateServiceCode)

The service.

savingsPlanOfferingRate_unit :: Lens' SavingsPlanOfferingRate (Maybe SavingsPlanRateUnit)

The unit.

savingsPlanOfferingRate_properties :: Lens' SavingsPlanOfferingRate (Maybe [SavingsPlanOfferingRateProperty])

The properties.

SavingsPlanOfferingRateFilterElement

data SavingsPlanOfferingRateFilterElement

Information about a filter.

See: newSavingsPlanOfferingRateFilterElement smart constructor.

Constructors

SavingsPlanOfferingRateFilterElement'

Fields

values :: Maybe [Text]

The filter values.

name :: Maybe SavingsPlanRateFilterAttribute

The filter name.

newSavingsPlanOfferingRateFilterElement :: SavingsPlanOfferingRateFilterElement

Create a value of SavingsPlanOfferingRateFilterElement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:SavingsPlanOfferingRateFilterElement', savingsPlanOfferingRateFilterElement_values - The filter values.

$sel:name:SavingsPlanOfferingRateFilterElement', savingsPlanOfferingRateFilterElement_name - The filter name.

savingsPlanOfferingRateFilterElement_values :: Lens' SavingsPlanOfferingRateFilterElement (Maybe [Text])

The filter values.

savingsPlanOfferingRateFilterElement_name :: Lens' SavingsPlanOfferingRateFilterElement (Maybe SavingsPlanRateFilterAttribute)

The filter name.

SavingsPlanOfferingRateProperty

data SavingsPlanOfferingRateProperty

Information about a property.

See: newSavingsPlanOfferingRateProperty smart constructor.

Constructors

SavingsPlanOfferingRateProperty'

Fields

value :: Maybe Text

The property value.

name :: Maybe Text

The property name.

newSavingsPlanOfferingRateProperty :: SavingsPlanOfferingRateProperty

Create a value of SavingsPlanOfferingRateProperty with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:SavingsPlanOfferingRateProperty', savingsPlanOfferingRateProperty_value - The property value.

$sel:name:SavingsPlanOfferingRateProperty', savingsPlanOfferingRateProperty_name - The property name.

savingsPlanOfferingRateProperty_value :: Lens' SavingsPlanOfferingRateProperty (Maybe Text)

The property value.

savingsPlanOfferingRateProperty_name :: Lens' SavingsPlanOfferingRateProperty (Maybe Text)

The property name.

SavingsPlanRate

data SavingsPlanRate

Information about a Savings Plan rate.

See: newSavingsPlanRate smart constructor.

Constructors

SavingsPlanRate'

Fields

operation :: Maybe Text

The specific AWS operation for the line item in the billing report.

usageType :: Maybe Text

The usage details of the line item in the billing report.

productType :: Maybe SavingsPlanProductType

The product type.

currency :: Maybe CurrencyCode

The currency.

rate :: Maybe Text

The rate.

serviceCode :: Maybe SavingsPlanRateServiceCode

The service.

unit :: Maybe SavingsPlanRateUnit

The unit.

properties :: Maybe [SavingsPlanRateProperty]

The properties.

newSavingsPlanRate :: SavingsPlanRate

Create a value of SavingsPlanRate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operation:SavingsPlanRate', savingsPlanRate_operation - The specific AWS operation for the line item in the billing report.

$sel:usageType:SavingsPlanRate', savingsPlanRate_usageType - The usage details of the line item in the billing report.

$sel:productType:SavingsPlanRate', savingsPlanRate_productType - The product type.

$sel:currency:SavingsPlanRate', savingsPlanRate_currency - The currency.

$sel:rate:SavingsPlanRate', savingsPlanRate_rate - The rate.

$sel:serviceCode:SavingsPlanRate', savingsPlanRate_serviceCode - The service.

$sel:unit:SavingsPlanRate', savingsPlanRate_unit - The unit.

$sel:properties:SavingsPlanRate', savingsPlanRate_properties - The properties.

savingsPlanRate_operation :: Lens' SavingsPlanRate (Maybe Text)

The specific AWS operation for the line item in the billing report.

savingsPlanRate_usageType :: Lens' SavingsPlanRate (Maybe Text)

The usage details of the line item in the billing report.

savingsPlanRate_productType :: Lens' SavingsPlanRate (Maybe SavingsPlanProductType)

The product type.

savingsPlanRate_currency :: Lens' SavingsPlanRate (Maybe CurrencyCode)

The currency.

savingsPlanRate_rate :: Lens' SavingsPlanRate (Maybe Text)

The rate.

savingsPlanRate_serviceCode :: Lens' SavingsPlanRate (Maybe SavingsPlanRateServiceCode)

The service.

savingsPlanRate_unit :: Lens' SavingsPlanRate (Maybe SavingsPlanRateUnit)

The unit.

savingsPlanRate_properties :: Lens' SavingsPlanRate (Maybe [SavingsPlanRateProperty])

The properties.

SavingsPlanRateFilter

data SavingsPlanRateFilter

Information about a filter.

See: newSavingsPlanRateFilter smart constructor.

Constructors

SavingsPlanRateFilter'

Fields

values :: Maybe [Text]

The filter values.

name :: Maybe SavingsPlanRateFilterName

The filter name.

newSavingsPlanRateFilter :: SavingsPlanRateFilter

Create a value of SavingsPlanRateFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:SavingsPlanRateFilter', savingsPlanRateFilter_values - The filter values.

$sel:name:SavingsPlanRateFilter', savingsPlanRateFilter_name - The filter name.

savingsPlanRateFilter_values :: Lens' SavingsPlanRateFilter (Maybe [Text])

The filter values.

savingsPlanRateFilter_name :: Lens' SavingsPlanRateFilter (Maybe SavingsPlanRateFilterName)

The filter name.

SavingsPlanRateProperty

data SavingsPlanRateProperty

Information about a property.

See: newSavingsPlanRateProperty smart constructor.

Constructors

SavingsPlanRateProperty'

Fields

value :: Maybe Text

The property value.

name :: Maybe SavingsPlanRatePropertyKey

The property name.

newSavingsPlanRateProperty :: SavingsPlanRateProperty

Create a value of SavingsPlanRateProperty with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:SavingsPlanRateProperty', savingsPlanRateProperty_value - The property value.

$sel:name:SavingsPlanRateProperty', savingsPlanRateProperty_name - The property name.

savingsPlanRateProperty_value :: Lens' SavingsPlanRateProperty (Maybe Text)

The property value.

savingsPlanRateProperty_name :: Lens' SavingsPlanRateProperty (Maybe SavingsPlanRatePropertyKey)

The property name.