Network.AWS.SavingsPlans.Types.SavingsPlanOfferingRate

Description

Documentation

data SavingsPlanOfferingRate

Information about a Savings Plan offering rate.

See: newSavingsPlanOfferingRate smart constructor.

Constructors

SavingsPlanOfferingRate'

Fields

operation :: Maybe Text

The specific AWS operation for the line item in the billing report.

usageType :: Maybe Text

The usage details of the line item in the billing report.

productType :: Maybe SavingsPlanProductType

The product type.

savingsPlanOffering :: Maybe ParentSavingsPlanOffering

The Savings Plan offering.

rate :: Maybe Text

The Savings Plan rate.

serviceCode :: Maybe SavingsPlanRateServiceCode

The service.

unit :: Maybe SavingsPlanRateUnit

The unit.

properties :: Maybe [SavingsPlanOfferingRateProperty]

The properties.

newSavingsPlanOfferingRate :: SavingsPlanOfferingRate

Create a value of SavingsPlanOfferingRate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operation:SavingsPlanOfferingRate', savingsPlanOfferingRate_operation - The specific AWS operation for the line item in the billing report.

$sel:usageType:SavingsPlanOfferingRate', savingsPlanOfferingRate_usageType - The usage details of the line item in the billing report.

$sel:productType:SavingsPlanOfferingRate', savingsPlanOfferingRate_productType - The product type.

$sel:savingsPlanOffering:SavingsPlanOfferingRate', savingsPlanOfferingRate_savingsPlanOffering - The Savings Plan offering.

$sel:rate:SavingsPlanOfferingRate', savingsPlanOfferingRate_rate - The Savings Plan rate.

$sel:serviceCode:SavingsPlanOfferingRate', savingsPlanOfferingRate_serviceCode - The service.

$sel:unit:SavingsPlanOfferingRate', savingsPlanOfferingRate_unit - The unit.

$sel:properties:SavingsPlanOfferingRate', savingsPlanOfferingRate_properties - The properties.

savingsPlanOfferingRate_operation :: Lens' SavingsPlanOfferingRate (Maybe Text)

The specific AWS operation for the line item in the billing report.

savingsPlanOfferingRate_usageType :: Lens' SavingsPlanOfferingRate (Maybe Text)

The usage details of the line item in the billing report.

savingsPlanOfferingRate_productType :: Lens' SavingsPlanOfferingRate (Maybe SavingsPlanProductType)

The product type.

savingsPlanOfferingRate_savingsPlanOffering :: Lens' SavingsPlanOfferingRate (Maybe ParentSavingsPlanOffering)

The Savings Plan offering.

savingsPlanOfferingRate_rate :: Lens' SavingsPlanOfferingRate (Maybe Text)

The Savings Plan rate.

savingsPlanOfferingRate_serviceCode :: Lens' SavingsPlanOfferingRate (Maybe SavingsPlanRateServiceCode)

The service.

savingsPlanOfferingRate_unit :: Lens' SavingsPlanOfferingRate (Maybe SavingsPlanRateUnit)

The unit.

savingsPlanOfferingRate_properties :: Lens' SavingsPlanOfferingRate (Maybe [SavingsPlanOfferingRateProperty])

The properties.