Network.AWS.Schemas.CreateSchema

Description

Creates a schema definition.

Inactive schemas will be deleted after two years.

Creating a Request

data CreateSchema

See: newCreateSchema smart constructor.

Constructors

CreateSchema'

Fields

description :: Maybe Text

A description of the schema.

tags :: Maybe (HashMap Text Text)

Tags associated with the schema.

registryName :: Text

The name of the registry.

schemaName :: Text

The name of the schema.

type' :: Type

The type of schema.

content :: Text

The source of the schema definition.

newCreateSchema

Arguments

:: Text
$sel:registryName:CreateSchema'

-> Text
$sel:schemaName:CreateSchema'

-> Type
$sel:type':CreateSchema'

-> Text
$sel:content:CreateSchema'

-> CreateSchema

Create a value of CreateSchema with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateSchema', createSchema_description - A description of the schema.

$sel:tags:CreateSchema', createSchema_tags - Tags associated with the schema.

$sel:registryName:CreateSchema', createSchema_registryName - The name of the registry.

$sel:schemaName:CreateSchema', createSchema_schemaName - The name of the schema.

$sel:type':CreateSchema', createSchema_type - The type of schema.

$sel:content:CreateSchema', createSchema_content - The source of the schema definition.

Request Lenses

createSchema_description :: Lens' CreateSchema (Maybe Text)

A description of the schema.

createSchema_tags :: Lens' CreateSchema (Maybe (HashMap Text Text))

Tags associated with the schema.

createSchema_registryName :: Lens' CreateSchema Text

The name of the registry.

createSchema_schemaName :: Lens' CreateSchema Text

The name of the schema.

createSchema_type :: Lens' CreateSchema Type

The type of schema.

createSchema_content :: Lens' CreateSchema Text

The source of the schema definition.

Destructuring the Response

data CreateSchemaResponse

See: newCreateSchemaResponse smart constructor.

Constructors

CreateSchemaResponse'

Fields

schemaVersion :: Maybe Text

The version number of the schema

schemaName :: Maybe Text

The name of the schema.

schemaArn :: Maybe Text

The ARN of the schema.

type' :: Maybe Text

The type of the schema.

lastModified :: Maybe POSIX

The date and time that schema was modified.

description :: Maybe Text

The description of the schema.

versionCreatedDate :: Maybe POSIX

The date the schema version was created.

tags :: Maybe (HashMap Text Text)

httpStatus :: Int

The response's http status code.

newCreateSchemaResponse

Arguments

:: Int
$sel:httpStatus:CreateSchemaResponse'

-> CreateSchemaResponse

Create a value of CreateSchemaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaVersion:CreateSchemaResponse', createSchemaResponse_schemaVersion - The version number of the schema

$sel:schemaName:CreateSchema', createSchemaResponse_schemaName - The name of the schema.

$sel:schemaArn:CreateSchemaResponse', createSchemaResponse_schemaArn - The ARN of the schema.

$sel:type':CreateSchema', createSchemaResponse_type - The type of the schema.

$sel:lastModified:CreateSchemaResponse', createSchemaResponse_lastModified - The date and time that schema was modified.

$sel:description:CreateSchema', createSchemaResponse_description - The description of the schema.

$sel:versionCreatedDate:CreateSchemaResponse', createSchemaResponse_versionCreatedDate - The date the schema version was created.

$sel:tags:CreateSchema', createSchemaResponse_tags - Undocumented member.

$sel:httpStatus:CreateSchemaResponse', createSchemaResponse_httpStatus - The response's http status code.

Response Lenses

createSchemaResponse_schemaVersion :: Lens' CreateSchemaResponse (Maybe Text)

The version number of the schema

createSchemaResponse_schemaName :: Lens' CreateSchemaResponse (Maybe Text)

The name of the schema.

createSchemaResponse_schemaArn :: Lens' CreateSchemaResponse (Maybe Text)

The ARN of the schema.

createSchemaResponse_type :: Lens' CreateSchemaResponse (Maybe Text)

The type of the schema.

createSchemaResponse_lastModified :: Lens' CreateSchemaResponse (Maybe UTCTime)

The date and time that schema was modified.

createSchemaResponse_description :: Lens' CreateSchemaResponse (Maybe Text)

The description of the schema.

createSchemaResponse_versionCreatedDate :: Lens' CreateSchemaResponse (Maybe UTCTime)

The date the schema version was created.

createSchemaResponse_tags :: Lens' CreateSchemaResponse (Maybe (HashMap Text Text))

Undocumented member.

createSchemaResponse_httpStatus :: Lens' CreateSchemaResponse Int

The response's http status code.