Network.AWS.Schemas.DescribeDiscoverer

Description

Describes the discoverer.

Creating a Request

data DescribeDiscoverer

See: newDescribeDiscoverer smart constructor.

Constructors

DescribeDiscoverer'

Fields

discovererId :: Text

The ID of the discoverer.

newDescribeDiscoverer

Arguments

:: Text
$sel:discovererId:DescribeDiscoverer'

-> DescribeDiscoverer

Create a value of DescribeDiscoverer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:discovererId:DescribeDiscoverer', describeDiscoverer_discovererId - The ID of the discoverer.

Request Lenses

describeDiscoverer_discovererId :: Lens' DescribeDiscoverer Text

The ID of the discoverer.

Destructuring the Response

data DescribeDiscovererResponse

See: newDescribeDiscovererResponse smart constructor.

Constructors

DescribeDiscovererResponse'

Fields

state :: Maybe DiscovererState

The state of the discoverer.

crossAccount :: Maybe Bool

The Status if the discoverer will discover schemas from events sent from another account.

sourceArn :: Maybe Text

The ARN of the event bus.

discovererId :: Maybe Text

The ID of the discoverer.

description :: Maybe Text

The description of the discoverer.

tags :: Maybe (HashMap Text Text)

Tags associated with the resource.

discovererArn :: Maybe Text

The ARN of the discoverer.

httpStatus :: Int

The response's http status code.

newDescribeDiscovererResponse

Arguments

:: Int
$sel:httpStatus:DescribeDiscovererResponse'

-> DescribeDiscovererResponse

Create a value of DescribeDiscovererResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:DescribeDiscovererResponse', describeDiscovererResponse_state - The state of the discoverer.

$sel:crossAccount:DescribeDiscovererResponse', describeDiscovererResponse_crossAccount - The Status if the discoverer will discover schemas from events sent from another account.

$sel:sourceArn:DescribeDiscovererResponse', describeDiscovererResponse_sourceArn - The ARN of the event bus.

$sel:discovererId:DescribeDiscoverer', describeDiscovererResponse_discovererId - The ID of the discoverer.

$sel:description:DescribeDiscovererResponse', describeDiscovererResponse_description - The description of the discoverer.

$sel:tags:DescribeDiscovererResponse', describeDiscovererResponse_tags - Tags associated with the resource.

$sel:discovererArn:DescribeDiscovererResponse', describeDiscovererResponse_discovererArn - The ARN of the discoverer.

$sel:httpStatus:DescribeDiscovererResponse', describeDiscovererResponse_httpStatus - The response's http status code.

Response Lenses

describeDiscovererResponse_state :: Lens' DescribeDiscovererResponse (Maybe DiscovererState)

The state of the discoverer.

describeDiscovererResponse_crossAccount :: Lens' DescribeDiscovererResponse (Maybe Bool)

The Status if the discoverer will discover schemas from events sent from another account.

describeDiscovererResponse_sourceArn :: Lens' DescribeDiscovererResponse (Maybe Text)

The ARN of the event bus.

describeDiscovererResponse_discovererId :: Lens' DescribeDiscovererResponse (Maybe Text)

The ID of the discoverer.

describeDiscovererResponse_description :: Lens' DescribeDiscovererResponse (Maybe Text)

The description of the discoverer.

describeDiscovererResponse_tags :: Lens' DescribeDiscovererResponse (Maybe (HashMap Text Text))

Tags associated with the resource.

describeDiscovererResponse_discovererArn :: Lens' DescribeDiscovererResponse (Maybe Text)

The ARN of the discoverer.

describeDiscovererResponse_httpStatus :: Lens' DescribeDiscovererResponse Int

The response's http status code.