Network.AWS.Schemas.GetDiscoveredSchema

Description

Get the discovered schema that was generated based on sampled events.

Creating a Request

data GetDiscoveredSchema

See: newGetDiscoveredSchema smart constructor.

Constructors

GetDiscoveredSchema'

Fields

type' :: Type

The type of event.

events :: NonEmpty Text

An array of strings where each string is a JSON event. These are the events that were used to generate the schema. The array includes a single type of event and has a maximum size of 10 events.

newGetDiscoveredSchema

Arguments

:: Type
$sel:type':GetDiscoveredSchema'

-> NonEmpty Text
$sel:events:GetDiscoveredSchema'

-> GetDiscoveredSchema

Create a value of GetDiscoveredSchema with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':GetDiscoveredSchema', getDiscoveredSchema_type - The type of event.

$sel:events:GetDiscoveredSchema', getDiscoveredSchema_events - An array of strings where each string is a JSON event. These are the events that were used to generate the schema. The array includes a single type of event and has a maximum size of 10 events.

Request Lenses

getDiscoveredSchema_type :: Lens' GetDiscoveredSchema Type

The type of event.

getDiscoveredSchema_events :: Lens' GetDiscoveredSchema (NonEmpty Text)

An array of strings where each string is a JSON event. These are the events that were used to generate the schema. The array includes a single type of event and has a maximum size of 10 events.

Destructuring the Response

data GetDiscoveredSchemaResponse

See: newGetDiscoveredSchemaResponse smart constructor.

Constructors

GetDiscoveredSchemaResponse'

Fields

content :: Maybe Text

The source of the schema definition.

httpStatus :: Int

The response's http status code.

newGetDiscoveredSchemaResponse

Arguments

:: Int
$sel:httpStatus:GetDiscoveredSchemaResponse'

-> GetDiscoveredSchemaResponse

Create a value of GetDiscoveredSchemaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:content:GetDiscoveredSchemaResponse', getDiscoveredSchemaResponse_content - The source of the schema definition.

$sel:httpStatus:GetDiscoveredSchemaResponse', getDiscoveredSchemaResponse_httpStatus - The response's http status code.

Response Lenses

getDiscoveredSchemaResponse_content :: Lens' GetDiscoveredSchemaResponse (Maybe Text)

The source of the schema definition.

getDiscoveredSchemaResponse_httpStatus :: Lens' GetDiscoveredSchemaResponse Int

The response's http status code.