Network.AWS.Schemas.Lens

Description

Operations

UpdateRegistry

updateRegistry_description :: Lens' UpdateRegistry (Maybe Text)

The description of the registry to update.

updateRegistry_registryName :: Lens' UpdateRegistry Text

The name of the registry.

updateRegistryResponse_registryName :: Lens' UpdateRegistryResponse (Maybe Text)

The name of the registry.

updateRegistryResponse_registryArn :: Lens' UpdateRegistryResponse (Maybe Text)

The ARN of the registry.

updateRegistryResponse_description :: Lens' UpdateRegistryResponse (Maybe Text)

The description of the registry.

updateRegistryResponse_tags :: Lens' UpdateRegistryResponse (Maybe (HashMap Text Text))

Tags associated with the registry.

updateRegistryResponse_httpStatus :: Lens' UpdateRegistryResponse Int

The response's http status code.

DeleteRegistry

deleteRegistry_registryName :: Lens' DeleteRegistry Text

The name of the registry.

SearchSchemas

searchSchemas_nextToken :: Lens' SearchSchemas (Maybe Text)

The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

searchSchemas_limit :: Lens' SearchSchemas (Maybe Int)

Undocumented member.

searchSchemas_registryName :: Lens' SearchSchemas Text

The name of the registry.

searchSchemas_keywords :: Lens' SearchSchemas Text

Specifying this limits the results to only schemas that include the provided keywords.

searchSchemasResponse_schemas :: Lens' SearchSchemasResponse (Maybe [SearchSchemaSummary])

An array of SearchSchemaSummary information.

searchSchemasResponse_nextToken :: Lens' SearchSchemasResponse (Maybe Text)

The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

searchSchemasResponse_httpStatus :: Lens' SearchSchemasResponse Int

The response's http status code.

StopDiscoverer

stopDiscoverer_discovererId :: Lens' StopDiscoverer Text

The ID of the discoverer.

stopDiscovererResponse_state :: Lens' StopDiscovererResponse (Maybe DiscovererState)

The state of the discoverer.

stopDiscovererResponse_discovererId :: Lens' StopDiscovererResponse (Maybe Text)

The ID of the discoverer.

stopDiscovererResponse_httpStatus :: Lens' StopDiscovererResponse Int

The response's http status code.

DeleteSchemaVersion

deleteSchemaVersion_schemaVersion :: Lens' DeleteSchemaVersion Text

The version number of the schema

deleteSchemaVersion_registryName :: Lens' DeleteSchemaVersion Text

The name of the registry.

deleteSchemaVersion_schemaName :: Lens' DeleteSchemaVersion Text

The name of the schema.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The ARN of the resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

Undocumented member.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

ListSchemaVersions

listSchemaVersions_nextToken :: Lens' ListSchemaVersions (Maybe Text)

The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

listSchemaVersions_limit :: Lens' ListSchemaVersions (Maybe Int)

Undocumented member.

listSchemaVersions_registryName :: Lens' ListSchemaVersions Text

The name of the registry.

listSchemaVersions_schemaName :: Lens' ListSchemaVersions Text

The name of the schema.

listSchemaVersionsResponse_schemaVersions :: Lens' ListSchemaVersionsResponse (Maybe [SchemaVersionSummary])

An array of schema version summaries.

listSchemaVersionsResponse_nextToken :: Lens' ListSchemaVersionsResponse (Maybe Text)

The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

listSchemaVersionsResponse_httpStatus :: Lens' ListSchemaVersionsResponse Int

The response's http status code.

ExportSchema

exportSchema_schemaVersion :: Lens' ExportSchema (Maybe Text)

Specifying this limits the results to only this schema version.

exportSchema_registryName :: Lens' ExportSchema Text

The name of the registry.

exportSchema_schemaName :: Lens' ExportSchema Text

The name of the schema.

exportSchema_type :: Lens' ExportSchema Text

Undocumented member.

exportSchemaResponse_schemaVersion :: Lens' ExportSchemaResponse (Maybe Text)

Undocumented member.

exportSchemaResponse_schemaName :: Lens' ExportSchemaResponse (Maybe Text)

Undocumented member.

exportSchemaResponse_content :: Lens' ExportSchemaResponse (Maybe Text)

Undocumented member.

exportSchemaResponse_schemaArn :: Lens' ExportSchemaResponse (Maybe Text)

Undocumented member.

exportSchemaResponse_type :: Lens' ExportSchemaResponse (Maybe Text)

Undocumented member.

exportSchemaResponse_httpStatus :: Lens' ExportSchemaResponse Int

The response's http status code.

GetDiscoveredSchema

getDiscoveredSchema_type :: Lens' GetDiscoveredSchema Type

The type of event.

getDiscoveredSchema_events :: Lens' GetDiscoveredSchema (NonEmpty Text)

An array of strings where each string is a JSON event. These are the events that were used to generate the schema. The array includes a single type of event and has a maximum size of 10 events.

getDiscoveredSchemaResponse_content :: Lens' GetDiscoveredSchemaResponse (Maybe Text)

The source of the schema definition.

getDiscoveredSchemaResponse_httpStatus :: Lens' GetDiscoveredSchemaResponse Int

The response's http status code.

DeleteDiscoverer

deleteDiscoverer_discovererId :: Lens' DeleteDiscoverer Text

The ID of the discoverer.

UpdateDiscoverer

updateDiscoverer_crossAccount :: Lens' UpdateDiscoverer (Maybe Bool)

Support discovery of schemas in events sent to the bus from another account. (default: true)

updateDiscoverer_description :: Lens' UpdateDiscoverer (Maybe Text)

The description of the discoverer to update.

updateDiscoverer_discovererId :: Lens' UpdateDiscoverer Text

The ID of the discoverer.

updateDiscovererResponse_state :: Lens' UpdateDiscovererResponse (Maybe DiscovererState)

The state of the discoverer.

updateDiscovererResponse_crossAccount :: Lens' UpdateDiscovererResponse (Maybe Bool)

The Status if the discoverer will discover schemas from events sent from another account.

updateDiscovererResponse_sourceArn :: Lens' UpdateDiscovererResponse (Maybe Text)

The ARN of the event bus.

updateDiscovererResponse_discovererId :: Lens' UpdateDiscovererResponse (Maybe Text)

The ID of the discoverer.

updateDiscovererResponse_description :: Lens' UpdateDiscovererResponse (Maybe Text)

The description of the discoverer.

updateDiscovererResponse_tags :: Lens' UpdateDiscovererResponse (Maybe (HashMap Text Text))

Tags associated with the resource.

updateDiscovererResponse_discovererArn :: Lens' UpdateDiscovererResponse (Maybe Text)

The ARN of the discoverer.

updateDiscovererResponse_httpStatus :: Lens' UpdateDiscovererResponse Int

The response's http status code.

ListDiscoverers

listDiscoverers_sourceArnPrefix :: Lens' ListDiscoverers (Maybe Text)

Specifying this limits the results to only those ARNs that start with the specified prefix.

listDiscoverers_discovererIdPrefix :: Lens' ListDiscoverers (Maybe Text)

Specifying this limits the results to only those discoverer IDs that start with the specified prefix.

listDiscoverers_nextToken :: Lens' ListDiscoverers (Maybe Text)

The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

listDiscoverers_limit :: Lens' ListDiscoverers (Maybe Int)

Undocumented member.

listDiscoverersResponse_discoverers :: Lens' ListDiscoverersResponse (Maybe [DiscovererSummary])

An array of DiscovererSummary information.

listDiscoverersResponse_nextToken :: Lens' ListDiscoverersResponse (Maybe Text)

The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

listDiscoverersResponse_httpStatus :: Lens' ListDiscoverersResponse Int

The response's http status code.

ListSchemas

listSchemas_schemaNamePrefix :: Lens' ListSchemas (Maybe Text)

Specifying this limits the results to only those schema names that start with the specified prefix.

listSchemas_nextToken :: Lens' ListSchemas (Maybe Text)

The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

listSchemas_limit :: Lens' ListSchemas (Maybe Int)

Undocumented member.

listSchemas_registryName :: Lens' ListSchemas Text

The name of the registry.

listSchemasResponse_schemas :: Lens' ListSchemasResponse (Maybe [SchemaSummary])

An array of schema summaries.

listSchemasResponse_nextToken :: Lens' ListSchemasResponse (Maybe Text)

The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

listSchemasResponse_httpStatus :: Lens' ListSchemasResponse Int

The response's http status code.

CreateDiscoverer

createDiscoverer_crossAccount :: Lens' CreateDiscoverer (Maybe Bool)

Support discovery of schemas in events sent to the bus from another account. (default: true).

createDiscoverer_description :: Lens' CreateDiscoverer (Maybe Text)

A description for the discoverer.

createDiscoverer_tags :: Lens' CreateDiscoverer (Maybe (HashMap Text Text))

Tags associated with the resource.

createDiscoverer_sourceArn :: Lens' CreateDiscoverer Text

The ARN of the event bus.

createDiscovererResponse_state :: Lens' CreateDiscovererResponse (Maybe DiscovererState)

The state of the discoverer.

createDiscovererResponse_crossAccount :: Lens' CreateDiscovererResponse (Maybe Bool)

The Status if the discoverer will discover schemas from events sent from another account.

createDiscovererResponse_sourceArn :: Lens' CreateDiscovererResponse (Maybe Text)

The ARN of the event bus.

createDiscovererResponse_discovererId :: Lens' CreateDiscovererResponse (Maybe Text)

The ID of the discoverer.

createDiscovererResponse_description :: Lens' CreateDiscovererResponse (Maybe Text)

The description of the discoverer.

createDiscovererResponse_tags :: Lens' CreateDiscovererResponse (Maybe (HashMap Text Text))

Tags associated with the resource.

createDiscovererResponse_discovererArn :: Lens' CreateDiscovererResponse (Maybe Text)

The ARN of the discoverer.

createDiscovererResponse_httpStatus :: Lens' CreateDiscovererResponse Int

The response's http status code.

DescribeRegistry

describeRegistry_registryName :: Lens' DescribeRegistry Text

The name of the registry.

describeRegistryResponse_registryName :: Lens' DescribeRegistryResponse (Maybe Text)

The name of the registry.

describeRegistryResponse_registryArn :: Lens' DescribeRegistryResponse (Maybe Text)

The ARN of the registry.

describeRegistryResponse_description :: Lens' DescribeRegistryResponse (Maybe Text)

The description of the registry.

describeRegistryResponse_tags :: Lens' DescribeRegistryResponse (Maybe (HashMap Text Text))

Tags associated with the registry.

describeRegistryResponse_httpStatus :: Lens' DescribeRegistryResponse Int

The response's http status code.

CreateRegistry

createRegistry_description :: Lens' CreateRegistry (Maybe Text)

A description of the registry to be created.

createRegistry_tags :: Lens' CreateRegistry (Maybe (HashMap Text Text))

Tags to associate with the registry.

createRegistry_registryName :: Lens' CreateRegistry Text

The name of the registry.

createRegistryResponse_registryName :: Lens' CreateRegistryResponse (Maybe Text)

The name of the registry.

createRegistryResponse_registryArn :: Lens' CreateRegistryResponse (Maybe Text)

The ARN of the registry.

createRegistryResponse_description :: Lens' CreateRegistryResponse (Maybe Text)

The description of the registry.

createRegistryResponse_tags :: Lens' CreateRegistryResponse (Maybe (HashMap Text Text))

Tags associated with the registry.

createRegistryResponse_httpStatus :: Lens' CreateRegistryResponse Int

The response's http status code.

ListRegistries

listRegistries_registryNamePrefix :: Lens' ListRegistries (Maybe Text)

Specifying this limits the results to only those registry names that start with the specified prefix.

listRegistries_nextToken :: Lens' ListRegistries (Maybe Text)

The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

listRegistries_scope :: Lens' ListRegistries (Maybe Text)

Can be set to Local or AWS to limit responses to your custom registries, or the ones provided by AWS.

listRegistries_limit :: Lens' ListRegistries (Maybe Int)

Undocumented member.

listRegistriesResponse_registries :: Lens' ListRegistriesResponse (Maybe [RegistrySummary])

An array of registry summaries.

listRegistriesResponse_nextToken :: Lens' ListRegistriesResponse (Maybe Text)

The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

listRegistriesResponse_httpStatus :: Lens' ListRegistriesResponse Int

The response's http status code.

DescribeDiscoverer

describeDiscoverer_discovererId :: Lens' DescribeDiscoverer Text

The ID of the discoverer.

describeDiscovererResponse_state :: Lens' DescribeDiscovererResponse (Maybe DiscovererState)

The state of the discoverer.

describeDiscovererResponse_crossAccount :: Lens' DescribeDiscovererResponse (Maybe Bool)

The Status if the discoverer will discover schemas from events sent from another account.

describeDiscovererResponse_sourceArn :: Lens' DescribeDiscovererResponse (Maybe Text)

The ARN of the event bus.

describeDiscovererResponse_discovererId :: Lens' DescribeDiscovererResponse (Maybe Text)

The ID of the discoverer.

describeDiscovererResponse_description :: Lens' DescribeDiscovererResponse (Maybe Text)

The description of the discoverer.

describeDiscovererResponse_tags :: Lens' DescribeDiscovererResponse (Maybe (HashMap Text Text))

Tags associated with the resource.

describeDiscovererResponse_discovererArn :: Lens' DescribeDiscovererResponse (Maybe Text)

The ARN of the discoverer.

describeDiscovererResponse_httpStatus :: Lens' DescribeDiscovererResponse Int

The response's http status code.

GetResourcePolicy

getResourcePolicy_registryName :: Lens' GetResourcePolicy (Maybe Text)

The name of the registry.

getResourcePolicyResponse_policy :: Lens' GetResourcePolicyResponse (Maybe Text)

The resource-based policy.

getResourcePolicyResponse_revisionId :: Lens' GetResourcePolicyResponse (Maybe Text)

The revision ID.

getResourcePolicyResponse_httpStatus :: Lens' GetResourcePolicyResponse Int

The response's http status code.

StartDiscoverer

startDiscoverer_discovererId :: Lens' StartDiscoverer Text

The ID of the discoverer.

startDiscovererResponse_state :: Lens' StartDiscovererResponse (Maybe DiscovererState)

The state of the discoverer.

startDiscovererResponse_discovererId :: Lens' StartDiscovererResponse (Maybe Text)

The ID of the discoverer.

startDiscovererResponse_httpStatus :: Lens' StartDiscovererResponse Int

The response's http status code.

DescribeSchema

describeSchema_schemaVersion :: Lens' DescribeSchema (Maybe Text)

Specifying this limits the results to only this schema version.

describeSchema_registryName :: Lens' DescribeSchema Text

The name of the registry.

describeSchema_schemaName :: Lens' DescribeSchema Text

The name of the schema.

describeSchemaResponse_schemaVersion :: Lens' DescribeSchemaResponse (Maybe Text)

The version number of the schema

describeSchemaResponse_schemaName :: Lens' DescribeSchemaResponse (Maybe Text)

The name of the schema.

describeSchemaResponse_content :: Lens' DescribeSchemaResponse (Maybe Text)

The source of the schema definition.

describeSchemaResponse_schemaArn :: Lens' DescribeSchemaResponse (Maybe Text)

The ARN of the schema.

describeSchemaResponse_type :: Lens' DescribeSchemaResponse (Maybe Text)

The type of the schema.

describeSchemaResponse_lastModified :: Lens' DescribeSchemaResponse (Maybe UTCTime)

The date and time that schema was modified.

describeSchemaResponse_description :: Lens' DescribeSchemaResponse (Maybe Text)

The description of the schema.

describeSchemaResponse_versionCreatedDate :: Lens' DescribeSchemaResponse (Maybe UTCTime)

The date the schema version was created.

describeSchemaResponse_tags :: Lens' DescribeSchemaResponse (Maybe (HashMap Text Text))

Tags associated with the resource.

describeSchemaResponse_httpStatus :: Lens' DescribeSchemaResponse Int

The response's http status code.

GetCodeBindingSource

getCodeBindingSource_schemaVersion :: Lens' GetCodeBindingSource (Maybe Text)

Specifying this limits the results to only this schema version.

getCodeBindingSource_registryName :: Lens' GetCodeBindingSource Text

The name of the registry.

getCodeBindingSource_schemaName :: Lens' GetCodeBindingSource Text

The name of the schema.

getCodeBindingSource_language :: Lens' GetCodeBindingSource Text

The language of the code binding.

getCodeBindingSourceResponse_body :: Lens' GetCodeBindingSourceResponse (Maybe ByteString)

Undocumented member.

getCodeBindingSourceResponse_httpStatus :: Lens' GetCodeBindingSourceResponse Int

The response's http status code.

PutCodeBinding

putCodeBinding_schemaVersion :: Lens' PutCodeBinding (Maybe Text)

Specifying this limits the results to only this schema version.

putCodeBinding_registryName :: Lens' PutCodeBinding Text

The name of the registry.

putCodeBinding_schemaName :: Lens' PutCodeBinding Text

The name of the schema.

putCodeBinding_language :: Lens' PutCodeBinding Text

The language of the code binding.

putCodeBindingResponse_status :: Lens' PutCodeBindingResponse (Maybe CodeGenerationStatus)

The current status of code binding generation.

putCodeBindingResponse_schemaVersion :: Lens' PutCodeBindingResponse (Maybe Text)

The version number of the schema.

putCodeBindingResponse_creationDate :: Lens' PutCodeBindingResponse (Maybe UTCTime)

The time and date that the code binding was created.

putCodeBindingResponse_lastModified :: Lens' PutCodeBindingResponse (Maybe UTCTime)

The date and time that code bindings were modified.

putCodeBindingResponse_httpStatus :: Lens' PutCodeBindingResponse Int

The response's http status code.

CreateSchema

createSchema_description :: Lens' CreateSchema (Maybe Text)

A description of the schema.

createSchema_tags :: Lens' CreateSchema (Maybe (HashMap Text Text))

Tags associated with the schema.

createSchema_registryName :: Lens' CreateSchema Text

The name of the registry.

createSchema_schemaName :: Lens' CreateSchema Text

The name of the schema.

createSchema_type :: Lens' CreateSchema Type

The type of schema.

createSchema_content :: Lens' CreateSchema Text

The source of the schema definition.

createSchemaResponse_schemaVersion :: Lens' CreateSchemaResponse (Maybe Text)

The version number of the schema

createSchemaResponse_schemaName :: Lens' CreateSchemaResponse (Maybe Text)

The name of the schema.

createSchemaResponse_schemaArn :: Lens' CreateSchemaResponse (Maybe Text)

The ARN of the schema.

createSchemaResponse_type :: Lens' CreateSchemaResponse (Maybe Text)

The type of the schema.

createSchemaResponse_lastModified :: Lens' CreateSchemaResponse (Maybe UTCTime)

The date and time that schema was modified.

createSchemaResponse_description :: Lens' CreateSchemaResponse (Maybe Text)

The description of the schema.

createSchemaResponse_versionCreatedDate :: Lens' CreateSchemaResponse (Maybe UTCTime)

The date the schema version was created.

createSchemaResponse_tags :: Lens' CreateSchemaResponse (Maybe (HashMap Text Text))

Undocumented member.

createSchemaResponse_httpStatus :: Lens' CreateSchemaResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The ARN of the resource.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

Tags associated with the resource.

UpdateSchema

updateSchema_content :: Lens' UpdateSchema (Maybe Text)

The source of the schema definition.

updateSchema_clientTokenId :: Lens' UpdateSchema (Maybe Text)

The ID of the client token.

updateSchema_type :: Lens' UpdateSchema (Maybe Type)

The schema type for the events schema.

updateSchema_description :: Lens' UpdateSchema (Maybe Text)

The description of the schema.

updateSchema_registryName :: Lens' UpdateSchema Text

The name of the registry.

updateSchema_schemaName :: Lens' UpdateSchema Text

The name of the schema.

updateSchemaResponse_schemaVersion :: Lens' UpdateSchemaResponse (Maybe Text)

The version number of the schema

updateSchemaResponse_schemaName :: Lens' UpdateSchemaResponse (Maybe Text)

The name of the schema.

updateSchemaResponse_schemaArn :: Lens' UpdateSchemaResponse (Maybe Text)

The ARN of the schema.

updateSchemaResponse_type :: Lens' UpdateSchemaResponse (Maybe Text)

The type of the schema.

updateSchemaResponse_lastModified :: Lens' UpdateSchemaResponse (Maybe UTCTime)

The date and time that schema was modified.

updateSchemaResponse_description :: Lens' UpdateSchemaResponse (Maybe Text)

The description of the schema.

updateSchemaResponse_versionCreatedDate :: Lens' UpdateSchemaResponse (Maybe UTCTime)

The date the schema version was created.

updateSchemaResponse_tags :: Lens' UpdateSchemaResponse (Maybe (HashMap Text Text))

Undocumented member.

updateSchemaResponse_httpStatus :: Lens' UpdateSchemaResponse Int

The response's http status code.

DeleteSchema

deleteSchema_registryName :: Lens' DeleteSchema Text

The name of the registry.

deleteSchema_schemaName :: Lens' DeleteSchema Text

The name of the schema.

PutResourcePolicy

putResourcePolicy_registryName :: Lens' PutResourcePolicy (Maybe Text)

The name of the registry.

putResourcePolicy_revisionId :: Lens' PutResourcePolicy (Maybe Text)

The revision ID of the policy.

putResourcePolicy_policy :: Lens' PutResourcePolicy Text

The resource-based policy.

putResourcePolicyResponse_policy :: Lens' PutResourcePolicyResponse (Maybe Text)

The resource-based policy.

putResourcePolicyResponse_revisionId :: Lens' PutResourcePolicyResponse (Maybe Text)

The revision ID of the policy.

putResourcePolicyResponse_httpStatus :: Lens' PutResourcePolicyResponse Int

The response's http status code.

DeleteResourcePolicy

deleteResourcePolicy_registryName :: Lens' DeleteResourcePolicy (Maybe Text)

The name of the registry.

UntagResource

untagResource_tagKeys :: Lens' UntagResource [Text]

Keys of key-value pairs.

untagResource_resourceArn :: Lens' UntagResource Text

The ARN of the resource.

DescribeCodeBinding

describeCodeBinding_schemaVersion :: Lens' DescribeCodeBinding (Maybe Text)

Specifying this limits the results to only this schema version.

describeCodeBinding_registryName :: Lens' DescribeCodeBinding Text

The name of the registry.

describeCodeBinding_schemaName :: Lens' DescribeCodeBinding Text

The name of the schema.

describeCodeBinding_language :: Lens' DescribeCodeBinding Text

The language of the code binding.

describeCodeBindingResponse_status :: Lens' DescribeCodeBindingResponse (Maybe CodeGenerationStatus)

The current status of code binding generation.

describeCodeBindingResponse_schemaVersion :: Lens' DescribeCodeBindingResponse (Maybe Text)

The version number of the schema.

describeCodeBindingResponse_creationDate :: Lens' DescribeCodeBindingResponse (Maybe UTCTime)

The time and date that the code binding was created.

describeCodeBindingResponse_lastModified :: Lens' DescribeCodeBindingResponse (Maybe UTCTime)

The date and time that code bindings were modified.

describeCodeBindingResponse_httpStatus :: Lens' DescribeCodeBindingResponse Int

The response's http status code.

Types

DiscovererSummary

discovererSummary_state :: Lens' DiscovererSummary (Maybe DiscovererState)

The state of the discoverer.

discovererSummary_crossAccount :: Lens' DiscovererSummary (Maybe Bool)

The Status if the discoverer will discover schemas from events sent from another account.

discovererSummary_sourceArn :: Lens' DiscovererSummary (Maybe Text)

The ARN of the event bus.

discovererSummary_discovererId :: Lens' DiscovererSummary (Maybe Text)

The ID of the discoverer.

discovererSummary_tags :: Lens' DiscovererSummary (Maybe (HashMap Text Text))

Tags associated with the resource.

discovererSummary_discovererArn :: Lens' DiscovererSummary (Maybe Text)

The ARN of the discoverer.

RegistrySummary

registrySummary_registryName :: Lens' RegistrySummary (Maybe Text)

The name of the registry.

registrySummary_registryArn :: Lens' RegistrySummary (Maybe Text)

The ARN of the registry.

registrySummary_tags :: Lens' RegistrySummary (Maybe (HashMap Text Text))

Tags associated with the registry.

SchemaSummary

schemaSummary_schemaName :: Lens' SchemaSummary (Maybe Text)

The name of the schema.

schemaSummary_schemaArn :: Lens' SchemaSummary (Maybe Text)

The ARN of the schema.

schemaSummary_lastModified :: Lens' SchemaSummary (Maybe UTCTime)

The date and time that schema was modified.

schemaSummary_tags :: Lens' SchemaSummary (Maybe (HashMap Text Text))

Tags associated with the schema.

schemaSummary_versionCount :: Lens' SchemaSummary (Maybe Integer)

The number of versions available for the schema.

SchemaVersionSummary

schemaVersionSummary_schemaVersion :: Lens' SchemaVersionSummary (Maybe Text)

The version number of the schema.

schemaVersionSummary_schemaName :: Lens' SchemaVersionSummary (Maybe Text)

The name of the schema.

schemaVersionSummary_schemaArn :: Lens' SchemaVersionSummary (Maybe Text)

The ARN of the schema version.

schemaVersionSummary_type :: Lens' SchemaVersionSummary (Maybe Type)

The type of schema.

SearchSchemaSummary

searchSchemaSummary_registryName :: Lens' SearchSchemaSummary (Maybe Text)

The name of the registry.

searchSchemaSummary_schemaVersions :: Lens' SearchSchemaSummary (Maybe [SearchSchemaVersionSummary])

An array of schema version summaries.

searchSchemaSummary_schemaName :: Lens' SearchSchemaSummary (Maybe Text)

The name of the schema.

searchSchemaSummary_schemaArn :: Lens' SearchSchemaSummary (Maybe Text)

The ARN of the schema.

SearchSchemaVersionSummary

searchSchemaVersionSummary_schemaVersion :: Lens' SearchSchemaVersionSummary (Maybe Text)

The version number of the schema

searchSchemaVersionSummary_createdDate :: Lens' SearchSchemaVersionSummary (Maybe UTCTime)

The date the schema version was created.

searchSchemaVersionSummary_type :: Lens' SearchSchemaVersionSummary (Maybe Type)

The type of schema.