Network.AWS.Schemas.ListRegistries

Description

List the registries.

This operation returns paginated results.

Creating a Request

data ListRegistries

See: newListRegistries smart constructor.

Constructors

ListRegistries'

Fields

registryNamePrefix :: Maybe Text

Specifying this limits the results to only those registry names that start with the specified prefix.

nextToken :: Maybe Text

The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

scope :: Maybe Text

Can be set to Local or AWS to limit responses to your custom registries, or the ones provided by AWS.

limit :: Maybe Int

newListRegistries :: ListRegistries

Create a value of ListRegistries with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryNamePrefix:ListRegistries', listRegistries_registryNamePrefix - Specifying this limits the results to only those registry names that start with the specified prefix.

$sel:nextToken:ListRegistries', listRegistries_nextToken - The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

$sel:scope:ListRegistries', listRegistries_scope - Can be set to Local or AWS to limit responses to your custom registries, or the ones provided by AWS.

$sel:limit:ListRegistries', listRegistries_limit - Undocumented member.

Request Lenses

listRegistries_registryNamePrefix :: Lens' ListRegistries (Maybe Text)

Specifying this limits the results to only those registry names that start with the specified prefix.

listRegistries_nextToken :: Lens' ListRegistries (Maybe Text)

The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

listRegistries_scope :: Lens' ListRegistries (Maybe Text)

Can be set to Local or AWS to limit responses to your custom registries, or the ones provided by AWS.

listRegistries_limit :: Lens' ListRegistries (Maybe Int)

Undocumented member.

Destructuring the Response

data ListRegistriesResponse

See: newListRegistriesResponse smart constructor.

Constructors

ListRegistriesResponse'

Fields

registries :: Maybe [RegistrySummary]

An array of registry summaries.

nextToken :: Maybe Text

The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

httpStatus :: Int

The response's http status code.

newListRegistriesResponse

Arguments

:: Int
$sel:httpStatus:ListRegistriesResponse'

-> ListRegistriesResponse

Create a value of ListRegistriesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registries:ListRegistriesResponse', listRegistriesResponse_registries - An array of registry summaries.

$sel:nextToken:ListRegistries', listRegistriesResponse_nextToken - The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

$sel:httpStatus:ListRegistriesResponse', listRegistriesResponse_httpStatus - The response's http status code.

Response Lenses

listRegistriesResponse_registries :: Lens' ListRegistriesResponse (Maybe [RegistrySummary])

An array of registry summaries.

listRegistriesResponse_nextToken :: Lens' ListRegistriesResponse (Maybe Text)

The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

listRegistriesResponse_httpStatus :: Lens' ListRegistriesResponse Int

The response's http status code.