Network.AWS.Schemas.Types

Description

Service Configuration

defaultService :: Service

API version 2019-12-02 of the Amazon Schemas SDK configuration.

Errors

_PreconditionFailedException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for PreconditionFailedException' errors.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for ConflictException' errors.

_ForbiddenException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for ForbiddenException' errors.

_GoneException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for GoneException' errors.

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for NotFoundException' errors.

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for TooManyRequestsException' errors.

_InternalServerErrorException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for InternalServerErrorException' errors.

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for ServiceUnavailableException' errors.

_UnauthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for UnauthorizedException' errors.

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for BadRequestException' errors.

CodeGenerationStatus

newtype CodeGenerationStatus

Constructors

CodeGenerationStatus'

Fields

fromCodeGenerationStatus :: Text

Bundled Patterns

pattern CodeGenerationStatus_CREATE_COMPLETE :: CodeGenerationStatus

pattern CodeGenerationStatus_CREATE_FAILED :: CodeGenerationStatus

pattern CodeGenerationStatus_CREATE_IN_PROGRESS :: CodeGenerationStatus

DiscovererState

newtype DiscovererState

Constructors

DiscovererState'

Fields

fromDiscovererState :: Text

Bundled Patterns

pattern DiscovererState_STARTED :: DiscovererState

pattern DiscovererState_STOPPED :: DiscovererState

Type

newtype Type

Constructors

Type'

Fields

fromType :: Text

Bundled Patterns

pattern Type_JSONSchemaDraft4 :: Type

pattern Type_OpenApi3 :: Type

DiscovererSummary

data DiscovererSummary

See: newDiscovererSummary smart constructor.

Constructors

DiscovererSummary'

Fields

state :: Maybe DiscovererState

The state of the discoverer.

crossAccount :: Maybe Bool

The Status if the discoverer will discover schemas from events sent from another account.

sourceArn :: Maybe Text

The ARN of the event bus.

discovererId :: Maybe Text

The ID of the discoverer.

tags :: Maybe (HashMap Text Text)

Tags associated with the resource.

discovererArn :: Maybe Text

The ARN of the discoverer.

newDiscovererSummary :: DiscovererSummary

Create a value of DiscovererSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:DiscovererSummary', discovererSummary_state - The state of the discoverer.

$sel:crossAccount:DiscovererSummary', discovererSummary_crossAccount - The Status if the discoverer will discover schemas from events sent from another account.

$sel:sourceArn:DiscovererSummary', discovererSummary_sourceArn - The ARN of the event bus.

$sel:discovererId:DiscovererSummary', discovererSummary_discovererId - The ID of the discoverer.

$sel:tags:DiscovererSummary', discovererSummary_tags - Tags associated with the resource.

$sel:discovererArn:DiscovererSummary', discovererSummary_discovererArn - The ARN of the discoverer.

discovererSummary_state :: Lens' DiscovererSummary (Maybe DiscovererState)

The state of the discoverer.

discovererSummary_crossAccount :: Lens' DiscovererSummary (Maybe Bool)

The Status if the discoverer will discover schemas from events sent from another account.

discovererSummary_sourceArn :: Lens' DiscovererSummary (Maybe Text)

The ARN of the event bus.

discovererSummary_discovererId :: Lens' DiscovererSummary (Maybe Text)

The ID of the discoverer.

discovererSummary_tags :: Lens' DiscovererSummary (Maybe (HashMap Text Text))

Tags associated with the resource.

discovererSummary_discovererArn :: Lens' DiscovererSummary (Maybe Text)

The ARN of the discoverer.

RegistrySummary

data RegistrySummary

See: newRegistrySummary smart constructor.

Constructors

RegistrySummary'

Fields

registryName :: Maybe Text

The name of the registry.

registryArn :: Maybe Text

The ARN of the registry.

tags :: Maybe (HashMap Text Text)

Tags associated with the registry.

newRegistrySummary :: RegistrySummary

Create a value of RegistrySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryName:RegistrySummary', registrySummary_registryName - The name of the registry.

$sel:registryArn:RegistrySummary', registrySummary_registryArn - The ARN of the registry.

$sel:tags:RegistrySummary', registrySummary_tags - Tags associated with the registry.

registrySummary_registryName :: Lens' RegistrySummary (Maybe Text)

The name of the registry.

registrySummary_registryArn :: Lens' RegistrySummary (Maybe Text)

The ARN of the registry.

registrySummary_tags :: Lens' RegistrySummary (Maybe (HashMap Text Text))

Tags associated with the registry.

SchemaSummary

data SchemaSummary

A summary of schema details.

See: newSchemaSummary smart constructor.

Constructors

SchemaSummary'

Fields

schemaName :: Maybe Text

The name of the schema.

schemaArn :: Maybe Text

The ARN of the schema.

lastModified :: Maybe POSIX

The date and time that schema was modified.

tags :: Maybe (HashMap Text Text)

Tags associated with the schema.

versionCount :: Maybe Integer

The number of versions available for the schema.

newSchemaSummary :: SchemaSummary

Create a value of SchemaSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaName:SchemaSummary', schemaSummary_schemaName - The name of the schema.

$sel:schemaArn:SchemaSummary', schemaSummary_schemaArn - The ARN of the schema.

$sel:lastModified:SchemaSummary', schemaSummary_lastModified - The date and time that schema was modified.

$sel:tags:SchemaSummary', schemaSummary_tags - Tags associated with the schema.

$sel:versionCount:SchemaSummary', schemaSummary_versionCount - The number of versions available for the schema.

schemaSummary_schemaName :: Lens' SchemaSummary (Maybe Text)

The name of the schema.

schemaSummary_schemaArn :: Lens' SchemaSummary (Maybe Text)

The ARN of the schema.

schemaSummary_lastModified :: Lens' SchemaSummary (Maybe UTCTime)

The date and time that schema was modified.

schemaSummary_tags :: Lens' SchemaSummary (Maybe (HashMap Text Text))

Tags associated with the schema.

schemaSummary_versionCount :: Lens' SchemaSummary (Maybe Integer)

The number of versions available for the schema.

SchemaVersionSummary

data SchemaVersionSummary

See: newSchemaVersionSummary smart constructor.

Constructors

SchemaVersionSummary'

Fields

schemaVersion :: Maybe Text

The version number of the schema.

schemaName :: Maybe Text

The name of the schema.

schemaArn :: Maybe Text

The ARN of the schema version.

type' :: Maybe Type

The type of schema.

newSchemaVersionSummary :: SchemaVersionSummary

Create a value of SchemaVersionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaVersion:SchemaVersionSummary', schemaVersionSummary_schemaVersion - The version number of the schema.

$sel:schemaName:SchemaVersionSummary', schemaVersionSummary_schemaName - The name of the schema.

$sel:schemaArn:SchemaVersionSummary', schemaVersionSummary_schemaArn - The ARN of the schema version.

$sel:type':SchemaVersionSummary', schemaVersionSummary_type - The type of schema.

schemaVersionSummary_schemaVersion :: Lens' SchemaVersionSummary (Maybe Text)

The version number of the schema.

schemaVersionSummary_schemaName :: Lens' SchemaVersionSummary (Maybe Text)

The name of the schema.

schemaVersionSummary_schemaArn :: Lens' SchemaVersionSummary (Maybe Text)

The ARN of the schema version.

schemaVersionSummary_type :: Lens' SchemaVersionSummary (Maybe Type)

The type of schema.

SearchSchemaSummary

data SearchSchemaSummary

See: newSearchSchemaSummary smart constructor.

Constructors

SearchSchemaSummary'

Fields

registryName :: Maybe Text

The name of the registry.

schemaVersions :: Maybe [SearchSchemaVersionSummary]

An array of schema version summaries.

schemaName :: Maybe Text

The name of the schema.

schemaArn :: Maybe Text

The ARN of the schema.

newSearchSchemaSummary :: SearchSchemaSummary

Create a value of SearchSchemaSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryName:SearchSchemaSummary', searchSchemaSummary_registryName - The name of the registry.

$sel:schemaVersions:SearchSchemaSummary', searchSchemaSummary_schemaVersions - An array of schema version summaries.

$sel:schemaName:SearchSchemaSummary', searchSchemaSummary_schemaName - The name of the schema.

$sel:schemaArn:SearchSchemaSummary', searchSchemaSummary_schemaArn - The ARN of the schema.

searchSchemaSummary_registryName :: Lens' SearchSchemaSummary (Maybe Text)

The name of the registry.

searchSchemaSummary_schemaVersions :: Lens' SearchSchemaSummary (Maybe [SearchSchemaVersionSummary])

An array of schema version summaries.

searchSchemaSummary_schemaName :: Lens' SearchSchemaSummary (Maybe Text)

The name of the schema.

searchSchemaSummary_schemaArn :: Lens' SearchSchemaSummary (Maybe Text)

The ARN of the schema.

SearchSchemaVersionSummary

data SearchSchemaVersionSummary

See: newSearchSchemaVersionSummary smart constructor.

Constructors

SearchSchemaVersionSummary'

Fields

schemaVersion :: Maybe Text

The version number of the schema

createdDate :: Maybe POSIX

The date the schema version was created.

type' :: Maybe Type

The type of schema.

newSearchSchemaVersionSummary :: SearchSchemaVersionSummary

Create a value of SearchSchemaVersionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaVersion:SearchSchemaVersionSummary', searchSchemaVersionSummary_schemaVersion - The version number of the schema

$sel:createdDate:SearchSchemaVersionSummary', searchSchemaVersionSummary_createdDate - The date the schema version was created.

$sel:type':SearchSchemaVersionSummary', searchSchemaVersionSummary_type - The type of schema.

searchSchemaVersionSummary_schemaVersion :: Lens' SearchSchemaVersionSummary (Maybe Text)

The version number of the schema

searchSchemaVersionSummary_createdDate :: Lens' SearchSchemaVersionSummary (Maybe UTCTime)

The date the schema version was created.

searchSchemaVersionSummary_type :: Lens' SearchSchemaVersionSummary (Maybe Type)

The type of schema.