Network.AWS.Schemas.UpdateRegistry

Description

Updates a registry.

Creating a Request

data UpdateRegistry

Updates the registry.

See: newUpdateRegistry smart constructor.

Constructors

UpdateRegistry'

Fields

description :: Maybe Text

The description of the registry to update.

registryName :: Text

The name of the registry.

newUpdateRegistry

Arguments

:: Text
$sel:registryName:UpdateRegistry'

-> UpdateRegistry

Create a value of UpdateRegistry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:UpdateRegistry', updateRegistry_description - The description of the registry to update.

$sel:registryName:UpdateRegistry', updateRegistry_registryName - The name of the registry.

Request Lenses

updateRegistry_description :: Lens' UpdateRegistry (Maybe Text)

The description of the registry to update.

updateRegistry_registryName :: Lens' UpdateRegistry Text

The name of the registry.

Destructuring the Response

data UpdateRegistryResponse

See: newUpdateRegistryResponse smart constructor.

Constructors

UpdateRegistryResponse'

Fields

registryName :: Maybe Text

The name of the registry.

registryArn :: Maybe Text

The ARN of the registry.

description :: Maybe Text

The description of the registry.

tags :: Maybe (HashMap Text Text)

Tags associated with the registry.

httpStatus :: Int

The response's http status code.

newUpdateRegistryResponse

Arguments

:: Int
$sel:httpStatus:UpdateRegistryResponse'

-> UpdateRegistryResponse

Create a value of UpdateRegistryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryName:UpdateRegistry', updateRegistryResponse_registryName - The name of the registry.

$sel:registryArn:UpdateRegistryResponse', updateRegistryResponse_registryArn - The ARN of the registry.

$sel:description:UpdateRegistry', updateRegistryResponse_description - The description of the registry.

$sel:tags:UpdateRegistryResponse', updateRegistryResponse_tags - Tags associated with the registry.

$sel:httpStatus:UpdateRegistryResponse', updateRegistryResponse_httpStatus - The response's http status code.

Response Lenses

updateRegistryResponse_registryName :: Lens' UpdateRegistryResponse (Maybe Text)

The name of the registry.

updateRegistryResponse_registryArn :: Lens' UpdateRegistryResponse (Maybe Text)

The ARN of the registry.

updateRegistryResponse_description :: Lens' UpdateRegistryResponse (Maybe Text)

The description of the registry.

updateRegistryResponse_tags :: Lens' UpdateRegistryResponse (Maybe (HashMap Text Text))

Tags associated with the registry.

updateRegistryResponse_httpStatus :: Lens' UpdateRegistryResponse Int

The response's http status code.