Network.AWS.SecretsManager.GetResourcePolicy

Description

Retrieves the JSON text of the resource-based policy document attached to the specified secret. The JSON request string input and response output displays formatted code with white space and line breaks for better readability. Submit your input as a single line JSON string.

Minimum permissions

To run this command, you must have the following permissions:

secretsmanager:GetResourcePolicy

Related operations

To attach a resource policy to a secret, use PutResourcePolicy.

To delete the resource-based policy attached to a secret, use DeleteResourcePolicy.

To list all of the currently available secrets, use ListSecrets.

Creating a Request

data GetResourcePolicy

See: newGetResourcePolicy smart constructor.

Constructors

GetResourcePolicy'

Fields

secretId :: Text

Specifies the secret that you want to retrieve the attached resource-based policy for. You can specify either the Amazon Resource Name (ARN) or the friendly name of the secret.

If you specify an ARN, we generally recommend that you specify a complete ARN. You can specify a partial ARN too—for example, if you don’t include the final hyphen and six random characters that Secrets Manager adds at the end of the ARN when you created the secret. A partial ARN match can work as long as it uniquely matches only one secret. However, if your secret has a name that ends in a hyphen followed by six characters (before Secrets Manager adds the hyphen and six characters to the ARN) and you try to use that as a partial ARN, then those characters cause Secrets Manager to assume that you’re specifying a complete ARN. This confusion can cause unexpected results. To avoid this situation, we recommend that you don’t create secret names ending with a hyphen followed by six characters.

If you specify an incomplete ARN without the random suffix, and instead provide the 'friendly name', you must not include the random suffix. If you do include the random suffix added by Secrets Manager, you receive either a ResourceNotFoundException or an AccessDeniedException error, depending on your permissions.

newGetResourcePolicy

Arguments

:: Text
$sel:secretId:GetResourcePolicy'

-> GetResourcePolicy

Create a value of GetResourcePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:secretId:GetResourcePolicy', getResourcePolicy_secretId - Specifies the secret that you want to retrieve the attached resource-based policy for. You can specify either the Amazon Resource Name (ARN) or the friendly name of the secret.

If you specify an ARN, we generally recommend that you specify a complete ARN. You can specify a partial ARN too—for example, if you don’t include the final hyphen and six random characters that Secrets Manager adds at the end of the ARN when you created the secret. A partial ARN match can work as long as it uniquely matches only one secret. However, if your secret has a name that ends in a hyphen followed by six characters (before Secrets Manager adds the hyphen and six characters to the ARN) and you try to use that as a partial ARN, then those characters cause Secrets Manager to assume that you’re specifying a complete ARN. This confusion can cause unexpected results. To avoid this situation, we recommend that you don’t create secret names ending with a hyphen followed by six characters.

If you specify an incomplete ARN without the random suffix, and instead provide the 'friendly name', you must not include the random suffix. If you do include the random suffix added by Secrets Manager, you receive either a ResourceNotFoundException or an AccessDeniedException error, depending on your permissions.

Request Lenses

getResourcePolicy_secretId :: Lens' GetResourcePolicy Text

Specifies the secret that you want to retrieve the attached resource-based policy for. You can specify either the Amazon Resource Name (ARN) or the friendly name of the secret.

If you specify an ARN, we generally recommend that you specify a complete ARN. You can specify a partial ARN too—for example, if you don’t include the final hyphen and six random characters that Secrets Manager adds at the end of the ARN when you created the secret. A partial ARN match can work as long as it uniquely matches only one secret. However, if your secret has a name that ends in a hyphen followed by six characters (before Secrets Manager adds the hyphen and six characters to the ARN) and you try to use that as a partial ARN, then those characters cause Secrets Manager to assume that you’re specifying a complete ARN. This confusion can cause unexpected results. To avoid this situation, we recommend that you don’t create secret names ending with a hyphen followed by six characters.

If you specify an incomplete ARN without the random suffix, and instead provide the 'friendly name', you must not include the random suffix. If you do include the random suffix added by Secrets Manager, you receive either a ResourceNotFoundException or an AccessDeniedException error, depending on your permissions.

Destructuring the Response

data GetResourcePolicyResponse

See: newGetResourcePolicyResponse smart constructor.

Constructors

GetResourcePolicyResponse'

Fields

arn :: Maybe Text

The ARN of the secret that the resource-based policy was retrieved for.

resourcePolicy :: Maybe Text

A JSON-formatted string that describes the permissions that are associated with the attached secret. These permissions are combined with any permissions that are associated with the user or role that attempts to access this secret. The combined permissions specify who can access the secret and what actions they can perform. For more information, see Authentication and Access Control for Amazon Web Services Secrets Manager in the Amazon Web Services Secrets Manager User Guide.

name :: Maybe Text

The friendly name of the secret that the resource-based policy was retrieved for.

httpStatus :: Int

The response's http status code.

newGetResourcePolicyResponse

Arguments

:: Int
$sel:httpStatus:GetResourcePolicyResponse'

-> GetResourcePolicyResponse

Create a value of GetResourcePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:GetResourcePolicyResponse', getResourcePolicyResponse_arn - The ARN of the secret that the resource-based policy was retrieved for.

$sel:resourcePolicy:GetResourcePolicyResponse', getResourcePolicyResponse_resourcePolicy - A JSON-formatted string that describes the permissions that are associated with the attached secret. These permissions are combined with any permissions that are associated with the user or role that attempts to access this secret. The combined permissions specify who can access the secret and what actions they can perform. For more information, see Authentication and Access Control for Amazon Web Services Secrets Manager in the Amazon Web Services Secrets Manager User Guide.

$sel:name:GetResourcePolicyResponse', getResourcePolicyResponse_name - The friendly name of the secret that the resource-based policy was retrieved for.

$sel:httpStatus:GetResourcePolicyResponse', getResourcePolicyResponse_httpStatus - The response's http status code.

Response Lenses

getResourcePolicyResponse_arn :: Lens' GetResourcePolicyResponse (Maybe Text)

The ARN of the secret that the resource-based policy was retrieved for.

getResourcePolicyResponse_resourcePolicy :: Lens' GetResourcePolicyResponse (Maybe Text)

A JSON-formatted string that describes the permissions that are associated with the attached secret. These permissions are combined with any permissions that are associated with the user or role that attempts to access this secret. The combined permissions specify who can access the secret and what actions they can perform. For more information, see Authentication and Access Control for Amazon Web Services Secrets Manager in the Amazon Web Services Secrets Manager User Guide.

getResourcePolicyResponse_name :: Lens' GetResourcePolicyResponse (Maybe Text)

The friendly name of the secret that the resource-based policy was retrieved for.

getResourcePolicyResponse_httpStatus :: Lens' GetResourcePolicyResponse Int

The response's http status code.