Network.AWS.SecretsManager.Types

Description

Service Configuration

defaultService :: Service

API version 2017-10-17 of the Amazon Secrets Manager SDK configuration.

Errors

_MalformedPolicyDocumentException :: AsError a => Getting (First ServiceError) a ServiceError

You provided a resource-based policy with syntax errors.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

You provided an invalid value for a parameter.

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

You provided a parameter value that is not valid for the current state of the resource.

Possible causes:

You tried to perform the operation on a secret that's currently marked deleted.

You tried to enable rotation on a secret that doesn't already have a Lambda function ARN configured and you didn't include such an ARN as a parameter in this call.

_DecryptionFailure :: AsError a => Getting (First ServiceError) a ServiceError

Secrets Manager can't decrypt the protected secret text using the provided KMS key.

_PublicPolicyException :: AsError a => Getting (First ServiceError) a ServiceError

The BlockPublicPolicy parameter is set to true and the resource policy did not prevent broad access to the secret.

_EncryptionFailure :: AsError a => Getting (First ServiceError) a ServiceError

Secrets Manager can't encrypt the protected secret text using the provided KMS key. Check that the customer master key (CMK) is available, enabled, and not in an invalid state. For more information, see How Key State Affects Use of a Customer Master Key.

_PreconditionNotMetException :: AsError a => Getting (First ServiceError) a ServiceError

The request failed because you did not complete all the prerequisite steps.

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

You provided an invalid NextToken value.

_InternalServiceError :: AsError a => Getting (First ServiceError) a ServiceError

An error occurred on the server side.

_ResourceExistsException :: AsError a => Getting (First ServiceError) a ServiceError

A resource with the ID you requested already exists.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

We can't find the resource that you asked for.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request failed because it would exceed one of the Secrets Manager internal limits.

FilterNameStringType

newtype FilterNameStringType

Constructors

FilterNameStringType'

Fields

fromFilterNameStringType :: Text

Bundled Patterns

pattern FilterNameStringType_All :: FilterNameStringType

pattern FilterNameStringType_Description :: FilterNameStringType

pattern FilterNameStringType_Name :: FilterNameStringType

pattern FilterNameStringType_Primary_region :: FilterNameStringType

pattern FilterNameStringType_Tag_key :: FilterNameStringType

pattern FilterNameStringType_Tag_value :: FilterNameStringType

SortOrderType

newtype SortOrderType

Constructors

SortOrderType'

Fields

fromSortOrderType :: Text

Bundled Patterns

pattern SortOrderType_Asc :: SortOrderType

pattern SortOrderType_Desc :: SortOrderType

StatusType

newtype StatusType

Constructors

StatusType'

Fields

fromStatusType :: Text

Bundled Patterns

pattern StatusType_Failed :: StatusType

pattern StatusType_InProgress :: StatusType

pattern StatusType_InSync :: StatusType

Filter

data Filter

Allows you to add filters when you use the search function in Secrets Manager.

See: newFilter smart constructor.

Constructors

Filter'

Fields

values :: Maybe (NonEmpty Text)

Filters your list of secrets by a specific value.

You can prefix your search value with an exclamation mark (!) in order to perform negation filters.

key :: Maybe FilterNameStringType

Filters your list of secrets by a specific key.

newFilter :: Filter

Create a value of Filter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:Filter', filter_values - Filters your list of secrets by a specific value.

You can prefix your search value with an exclamation mark (!) in order to perform negation filters.

$sel:key:Filter', filter_key - Filters your list of secrets by a specific key.

filter_values :: Lens' Filter (Maybe (NonEmpty Text))

Filters your list of secrets by a specific value.

You can prefix your search value with an exclamation mark (!) in order to perform negation filters.

filter_key :: Lens' Filter (Maybe FilterNameStringType)

Filters your list of secrets by a specific key.

ReplicaRegionType

data ReplicaRegionType

(Optional) Custom type consisting of a Region (required) and the KmsKeyId which can be an ARN, Key ID, or Alias.

See: newReplicaRegionType smart constructor.

Constructors

ReplicaRegionType'

Fields

kmsKeyId :: Maybe Text

Can be an ARN, Key ID, or Alias.

region :: Maybe Text

Describes a single instance of Region objects.

newReplicaRegionType :: ReplicaRegionType

Create a value of ReplicaRegionType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyId:ReplicaRegionType', replicaRegionType_kmsKeyId - Can be an ARN, Key ID, or Alias.

$sel:region:ReplicaRegionType', replicaRegionType_region - Describes a single instance of Region objects.

replicaRegionType_kmsKeyId :: Lens' ReplicaRegionType (Maybe Text)

Can be an ARN, Key ID, or Alias.

replicaRegionType_region :: Lens' ReplicaRegionType (Maybe Text)

Describes a single instance of Region objects.

ReplicationStatusType

data ReplicationStatusType

A replication object consisting of a RegionReplicationStatus object and includes a Region, KMSKeyId, status, and status message.

See: newReplicationStatusType smart constructor.

Constructors

ReplicationStatusType'

Fields

status :: Maybe StatusType

The status can be InProgress, Failed, or InSync.

kmsKeyId :: Maybe Text

Can be an ARN, Key ID, or Alias.

statusMessage :: Maybe Text

Status message such as "/Secret with this name already exists in this region/".

region :: Maybe Text

The Region where replication occurs.

lastAccessedDate :: Maybe POSIX

The date that you last accessed the secret in the Region.

newReplicationStatusType :: ReplicationStatusType

Create a value of ReplicationStatusType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ReplicationStatusType', replicationStatusType_status - The status can be InProgress, Failed, or InSync.

$sel:kmsKeyId:ReplicationStatusType', replicationStatusType_kmsKeyId - Can be an ARN, Key ID, or Alias.

$sel:statusMessage:ReplicationStatusType', replicationStatusType_statusMessage - Status message such as "/Secret with this name already exists in this region/".

$sel:region:ReplicationStatusType', replicationStatusType_region - The Region where replication occurs.

$sel:lastAccessedDate:ReplicationStatusType', replicationStatusType_lastAccessedDate - The date that you last accessed the secret in the Region.

replicationStatusType_status :: Lens' ReplicationStatusType (Maybe StatusType)

The status can be InProgress, Failed, or InSync.

replicationStatusType_kmsKeyId :: Lens' ReplicationStatusType (Maybe Text)

Can be an ARN, Key ID, or Alias.

replicationStatusType_statusMessage :: Lens' ReplicationStatusType (Maybe Text)

Status message such as "/Secret with this name already exists in this region/".

replicationStatusType_region :: Lens' ReplicationStatusType (Maybe Text)

The Region where replication occurs.

replicationStatusType_lastAccessedDate :: Lens' ReplicationStatusType (Maybe UTCTime)

The date that you last accessed the secret in the Region.

RotationRulesType

data RotationRulesType

A structure that defines the rotation configuration for the secret.

See: newRotationRulesType smart constructor.

Constructors

RotationRulesType'

Fields

automaticallyAfterDays :: Maybe Natural

Specifies the number of days between automatic scheduled rotations of the secret.

Secrets Manager schedules the next rotation when the previous one is complete. Secrets Manager schedules the date by adding the rotation interval (number of days) to the actual date of the last rotation. The service chooses the hour within that 24-hour date window randomly. The minute is also chosen somewhat randomly, but weighted towards the top of the hour and influenced by a variety of factors that help distribute load.

newRotationRulesType :: RotationRulesType

Create a value of RotationRulesType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:automaticallyAfterDays:RotationRulesType', rotationRulesType_automaticallyAfterDays - Specifies the number of days between automatic scheduled rotations of the secret.

Secrets Manager schedules the next rotation when the previous one is complete. Secrets Manager schedules the date by adding the rotation interval (number of days) to the actual date of the last rotation. The service chooses the hour within that 24-hour date window randomly. The minute is also chosen somewhat randomly, but weighted towards the top of the hour and influenced by a variety of factors that help distribute load.

rotationRulesType_automaticallyAfterDays :: Lens' RotationRulesType (Maybe Natural)

Specifies the number of days between automatic scheduled rotations of the secret.

Secrets Manager schedules the next rotation when the previous one is complete. Secrets Manager schedules the date by adding the rotation interval (number of days) to the actual date of the last rotation. The service chooses the hour within that 24-hour date window randomly. The minute is also chosen somewhat randomly, but weighted towards the top of the hour and influenced by a variety of factors that help distribute load.

SecretListEntry

data SecretListEntry

A structure that contains the details about a secret. It does not include the encrypted SecretString and SecretBinary values. To get those values, use the GetSecretValue operation.

See: newSecretListEntry smart constructor.

Constructors

SecretListEntry'

Fields

lastChangedDate :: Maybe POSIX

The last date and time that this secret was modified in any way.

primaryRegion :: Maybe Text

The Region where Secrets Manager originated the secret.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the secret.

For more information about ARNs in Secrets Manager, see Policy Resources in the Amazon Web Services Secrets Manager User Guide.

secretVersionsToStages :: Maybe (HashMap Text (NonEmpty Text))

A list of all of the currently assigned SecretVersionStage staging labels and the SecretVersionId attached to each one. Staging labels are used to keep track of the different versions during the rotation process.

A version that does not have any SecretVersionStage is considered deprecated and subject to deletion. Such versions are not included in this list.

rotationRules :: Maybe RotationRulesType

A structure that defines the rotation configuration for the secret.

deletedDate :: Maybe POSIX

The date and time the deletion of the secret occurred. Not present on active secrets. The secret can be recovered until the number of days in the recovery window has passed, as specified in the RecoveryWindowInDays parameter of the DeleteSecret operation.

rotationEnabled :: Maybe Bool

Indicates whether automatic, scheduled rotation is enabled for this secret.

createdDate :: Maybe POSIX

The date and time when a secret was created.

kmsKeyId :: Maybe Text

The ARN or alias of the Amazon Web Services KMS customer master key (CMK) used to encrypt the SecretString and SecretBinary fields in each version of the secret. If you don't provide a key, then Secrets Manager defaults to encrypting the secret fields with the default KMS CMK, the key named awssecretsmanager, for this account.

name :: Maybe Text

The friendly name of the secret. You can use forward slashes in the name to represent a path hierarchy. For example, /prod/databases/dbserver1 could represent the secret for a server named dbserver1 in the folder databases in the folder prod.

owningService :: Maybe Text

Returns the name of the service that created the secret.

lastRotatedDate :: Maybe POSIX

The most recent date and time that the Secrets Manager rotation process was successfully completed. This value is null if the secret hasn't ever rotated.

lastAccessedDate :: Maybe POSIX

The last date that this secret was accessed. This value is truncated to midnight of the date and therefore shows only the date, not the time.

description :: Maybe Text

The user-provided description of the secret.

rotationLambdaARN :: Maybe Text

The ARN of an Amazon Web Services Lambda function invoked by Secrets Manager to rotate and expire the secret either automatically per the schedule or manually by a call to RotateSecret.

tags :: Maybe [Tag]

The list of user-defined tags associated with the secret. To add tags to a secret, use TagResource. To remove tags, use UntagResource.

newSecretListEntry :: SecretListEntry

Create a value of SecretListEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastChangedDate:SecretListEntry', secretListEntry_lastChangedDate - The last date and time that this secret was modified in any way.

$sel:primaryRegion:SecretListEntry', secretListEntry_primaryRegion - The Region where Secrets Manager originated the secret.

$sel:arn:SecretListEntry', secretListEntry_arn - The Amazon Resource Name (ARN) of the secret.

For more information about ARNs in Secrets Manager, see Policy Resources in the Amazon Web Services Secrets Manager User Guide.

$sel:secretVersionsToStages:SecretListEntry', secretListEntry_secretVersionsToStages - A list of all of the currently assigned SecretVersionStage staging labels and the SecretVersionId attached to each one. Staging labels are used to keep track of the different versions during the rotation process.

A version that does not have any SecretVersionStage is considered deprecated and subject to deletion. Such versions are not included in this list.

$sel:rotationRules:SecretListEntry', secretListEntry_rotationRules - A structure that defines the rotation configuration for the secret.

$sel:deletedDate:SecretListEntry', secretListEntry_deletedDate - The date and time the deletion of the secret occurred. Not present on active secrets. The secret can be recovered until the number of days in the recovery window has passed, as specified in the RecoveryWindowInDays parameter of the DeleteSecret operation.

$sel:rotationEnabled:SecretListEntry', secretListEntry_rotationEnabled - Indicates whether automatic, scheduled rotation is enabled for this secret.

$sel:createdDate:SecretListEntry', secretListEntry_createdDate - The date and time when a secret was created.

$sel:kmsKeyId:SecretListEntry', secretListEntry_kmsKeyId - The ARN or alias of the Amazon Web Services KMS customer master key (CMK) used to encrypt the SecretString and SecretBinary fields in each version of the secret. If you don't provide a key, then Secrets Manager defaults to encrypting the secret fields with the default KMS CMK, the key named awssecretsmanager, for this account.

$sel:name:SecretListEntry', secretListEntry_name - The friendly name of the secret. You can use forward slashes in the name to represent a path hierarchy. For example, /prod/databases/dbserver1 could represent the secret for a server named dbserver1 in the folder databases in the folder prod.

$sel:owningService:SecretListEntry', secretListEntry_owningService - Returns the name of the service that created the secret.

$sel:lastRotatedDate:SecretListEntry', secretListEntry_lastRotatedDate - The most recent date and time that the Secrets Manager rotation process was successfully completed. This value is null if the secret hasn't ever rotated.

$sel:lastAccessedDate:SecretListEntry', secretListEntry_lastAccessedDate - The last date that this secret was accessed. This value is truncated to midnight of the date and therefore shows only the date, not the time.

$sel:description:SecretListEntry', secretListEntry_description - The user-provided description of the secret.

$sel:rotationLambdaARN:SecretListEntry', secretListEntry_rotationLambdaARN - The ARN of an Amazon Web Services Lambda function invoked by Secrets Manager to rotate and expire the secret either automatically per the schedule or manually by a call to RotateSecret.

$sel:tags:SecretListEntry', secretListEntry_tags - The list of user-defined tags associated with the secret. To add tags to a secret, use TagResource. To remove tags, use UntagResource.

secretListEntry_lastChangedDate :: Lens' SecretListEntry (Maybe UTCTime)

The last date and time that this secret was modified in any way.

secretListEntry_primaryRegion :: Lens' SecretListEntry (Maybe Text)

The Region where Secrets Manager originated the secret.

secretListEntry_arn :: Lens' SecretListEntry (Maybe Text)

The Amazon Resource Name (ARN) of the secret.

For more information about ARNs in Secrets Manager, see Policy Resources in the Amazon Web Services Secrets Manager User Guide.

secretListEntry_secretVersionsToStages :: Lens' SecretListEntry (Maybe (HashMap Text (NonEmpty Text)))

A list of all of the currently assigned SecretVersionStage staging labels and the SecretVersionId attached to each one. Staging labels are used to keep track of the different versions during the rotation process.

A version that does not have any SecretVersionStage is considered deprecated and subject to deletion. Such versions are not included in this list.

secretListEntry_rotationRules :: Lens' SecretListEntry (Maybe RotationRulesType)

A structure that defines the rotation configuration for the secret.

secretListEntry_deletedDate :: Lens' SecretListEntry (Maybe UTCTime)

The date and time the deletion of the secret occurred. Not present on active secrets. The secret can be recovered until the number of days in the recovery window has passed, as specified in the RecoveryWindowInDays parameter of the DeleteSecret operation.

secretListEntry_rotationEnabled :: Lens' SecretListEntry (Maybe Bool)

Indicates whether automatic, scheduled rotation is enabled for this secret.

secretListEntry_createdDate :: Lens' SecretListEntry (Maybe UTCTime)

The date and time when a secret was created.

secretListEntry_kmsKeyId :: Lens' SecretListEntry (Maybe Text)

The ARN or alias of the Amazon Web Services KMS customer master key (CMK) used to encrypt the SecretString and SecretBinary fields in each version of the secret. If you don't provide a key, then Secrets Manager defaults to encrypting the secret fields with the default KMS CMK, the key named awssecretsmanager, for this account.

secretListEntry_name :: Lens' SecretListEntry (Maybe Text)

The friendly name of the secret. You can use forward slashes in the name to represent a path hierarchy. For example, /prod/databases/dbserver1 could represent the secret for a server named dbserver1 in the folder databases in the folder prod.

secretListEntry_owningService :: Lens' SecretListEntry (Maybe Text)

Returns the name of the service that created the secret.

secretListEntry_lastRotatedDate :: Lens' SecretListEntry (Maybe UTCTime)

The most recent date and time that the Secrets Manager rotation process was successfully completed. This value is null if the secret hasn't ever rotated.

secretListEntry_lastAccessedDate :: Lens' SecretListEntry (Maybe UTCTime)

The last date that this secret was accessed. This value is truncated to midnight of the date and therefore shows only the date, not the time.

secretListEntry_description :: Lens' SecretListEntry (Maybe Text)

The user-provided description of the secret.

secretListEntry_rotationLambdaARN :: Lens' SecretListEntry (Maybe Text)

The ARN of an Amazon Web Services Lambda function invoked by Secrets Manager to rotate and expire the secret either automatically per the schedule or manually by a call to RotateSecret.

secretListEntry_tags :: Lens' SecretListEntry (Maybe [Tag])

The list of user-defined tags associated with the secret. To add tags to a secret, use TagResource. To remove tags, use UntagResource.

SecretVersionsListEntry

data SecretVersionsListEntry

A structure that contains information about one version of a secret.

See: newSecretVersionsListEntry smart constructor.

Constructors

SecretVersionsListEntry'

Fields

versionId :: Maybe Text

The unique version identifier of this version of the secret.

versionStages :: Maybe (NonEmpty Text)

An array of staging labels that are currently associated with this version of the secret.

createdDate :: Maybe POSIX

The date and time this version of the secret was created.

kmsKeyIds :: Maybe [Text]

The KMS keys used to encrypt the secret version.

lastAccessedDate :: Maybe POSIX

The date that this version of the secret was last accessed. Note that the resolution of this field is at the date level and does not include the time.

newSecretVersionsListEntry :: SecretVersionsListEntry

Create a value of SecretVersionsListEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:SecretVersionsListEntry', secretVersionsListEntry_versionId - The unique version identifier of this version of the secret.

$sel:versionStages:SecretVersionsListEntry', secretVersionsListEntry_versionStages - An array of staging labels that are currently associated with this version of the secret.

$sel:createdDate:SecretVersionsListEntry', secretVersionsListEntry_createdDate - The date and time this version of the secret was created.

$sel:kmsKeyIds:SecretVersionsListEntry', secretVersionsListEntry_kmsKeyIds - The KMS keys used to encrypt the secret version.

$sel:lastAccessedDate:SecretVersionsListEntry', secretVersionsListEntry_lastAccessedDate - The date that this version of the secret was last accessed. Note that the resolution of this field is at the date level and does not include the time.

secretVersionsListEntry_versionId :: Lens' SecretVersionsListEntry (Maybe Text)

The unique version identifier of this version of the secret.

secretVersionsListEntry_versionStages :: Lens' SecretVersionsListEntry (Maybe (NonEmpty Text))

An array of staging labels that are currently associated with this version of the secret.

secretVersionsListEntry_createdDate :: Lens' SecretVersionsListEntry (Maybe UTCTime)

The date and time this version of the secret was created.

secretVersionsListEntry_kmsKeyIds :: Lens' SecretVersionsListEntry (Maybe [Text])

The KMS keys used to encrypt the secret version.

secretVersionsListEntry_lastAccessedDate :: Lens' SecretVersionsListEntry (Maybe UTCTime)

The date that this version of the secret was last accessed. Note that the resolution of this field is at the date level and does not include the time.

Tag

data Tag

A structure that contains information about a tag.

See: newTag smart constructor.

Constructors

Tag'

Fields

value :: Maybe Text

The string value associated with the key of the tag.

key :: Maybe Text

The key identifier, or name, of the tag.

newTag :: Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:Tag', tag_value - The string value associated with the key of the tag.

$sel:key:Tag', tag_key - The key identifier, or name, of the tag.

tag_value :: Lens' Tag (Maybe Text)

The string value associated with the key of the tag.

tag_key :: Lens' Tag (Maybe Text)

The key identifier, or name, of the tag.

ValidationErrorsEntry

data ValidationErrorsEntry

Displays errors that occurred during validation of the resource policy.

See: newValidationErrorsEntry smart constructor.

Constructors

ValidationErrorsEntry'

Fields

checkName :: Maybe Text

Checks the name of the policy.

errorMessage :: Maybe Text

Displays error messages if validation encounters problems during validation of the resource policy.

newValidationErrorsEntry :: ValidationErrorsEntry

Create a value of ValidationErrorsEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checkName:ValidationErrorsEntry', validationErrorsEntry_checkName - Checks the name of the policy.

$sel:errorMessage:ValidationErrorsEntry', validationErrorsEntry_errorMessage - Displays error messages if validation encounters problems during validation of the resource policy.

validationErrorsEntry_checkName :: Lens' ValidationErrorsEntry (Maybe Text)

Checks the name of the policy.

validationErrorsEntry_errorMessage :: Lens' ValidationErrorsEntry (Maybe Text)

Displays error messages if validation encounters problems during validation of the resource policy.