Network.AWS.SecretsManager.Types.SecretVersionsListEntry

Description

Documentation

data SecretVersionsListEntry

A structure that contains information about one version of a secret.

See: newSecretVersionsListEntry smart constructor.

Constructors

SecretVersionsListEntry'

Fields

createdDate :: Maybe POSIX

The date and time this version of the secret was created.

kmsKeyIds :: Maybe [Text]

The KMS keys used to encrypt the secret version.

versionStages :: Maybe (NonEmpty Text)

An array of staging labels that are currently associated with this version of the secret.

versionId :: Maybe Text

The unique version identifier of this version of the secret.

lastAccessedDate :: Maybe POSIX

The date that this version of the secret was last accessed. Note that the resolution of this field is at the date level and does not include the time.

newSecretVersionsListEntry :: SecretVersionsListEntry

Create a value of SecretVersionsListEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdDate:SecretVersionsListEntry', secretVersionsListEntry_createdDate - The date and time this version of the secret was created.

$sel:kmsKeyIds:SecretVersionsListEntry', secretVersionsListEntry_kmsKeyIds - The KMS keys used to encrypt the secret version.

$sel:versionStages:SecretVersionsListEntry', secretVersionsListEntry_versionStages - An array of staging labels that are currently associated with this version of the secret.

$sel:versionId:SecretVersionsListEntry', secretVersionsListEntry_versionId - The unique version identifier of this version of the secret.

$sel:lastAccessedDate:SecretVersionsListEntry', secretVersionsListEntry_lastAccessedDate - The date that this version of the secret was last accessed. Note that the resolution of this field is at the date level and does not include the time.

secretVersionsListEntry_createdDate :: Lens' SecretVersionsListEntry (Maybe UTCTime)

The date and time this version of the secret was created.

secretVersionsListEntry_kmsKeyIds :: Lens' SecretVersionsListEntry (Maybe [Text])

The KMS keys used to encrypt the secret version.

secretVersionsListEntry_versionStages :: Lens' SecretVersionsListEntry (Maybe (NonEmpty Text))

An array of staging labels that are currently associated with this version of the secret.

secretVersionsListEntry_versionId :: Lens' SecretVersionsListEntry (Maybe Text)

The unique version identifier of this version of the secret.

secretVersionsListEntry_lastAccessedDate :: Lens' SecretVersionsListEntry (Maybe UTCTime)

The date that this version of the secret was last accessed. Note that the resolution of this field is at the date level and does not include the time.