Network.AWS.SecretsManager.UpdateSecretVersionStage

Description

Modifies the staging labels attached to a version of a secret. Staging labels are used to track a version as it progresses through the secret rotation process. You can attach a staging label to only one version of a secret at a time. If a staging label to be added is already attached to another version, then it is moved--removed from the other version first and then attached to this one. For more information about staging labels, see Staging Labels in the Amazon Web Services Secrets Manager User Guide.

The staging labels that you specify in the VersionStage parameter are added to the existing list of staging labels--they don't replace it.

You can move the AWSCURRENT staging label to this version by including it in this call.

Whenever you move AWSCURRENT, Secrets Manager automatically moves the label AWSPREVIOUS to the version that AWSCURRENT was removed from.

If this action results in the last label being removed from a version, then the version is considered to be 'deprecated' and can be deleted by Secrets Manager.

Minimum permissions

To run this command, you must have the following permissions:

secretsmanager:UpdateSecretVersionStage

Related operations

To get the list of staging labels that are currently associated with a version of a secret, use DescribeSecret and examine the SecretVersionsToStages response value.

Creating a Request

data UpdateSecretVersionStage

See: newUpdateSecretVersionStage smart constructor.

Constructors

UpdateSecretVersionStage'

Fields

removeFromVersionId :: Maybe Text

Specifies the secret version ID of the version that the staging label is to be removed from. If the staging label you are trying to attach to one version is already attached to a different version, then you must include this parameter and specify the version that the label is to be removed from. If the label is attached and you either do not specify this parameter, or the version ID does not match, then the operation fails.

moveToVersionId :: Maybe Text

(Optional) The secret version ID that you want to add the staging label. If you want to remove a label from a version, then do not specify this parameter.

If the staging label is already attached to a different version of the secret, then you must also specify the RemoveFromVersionId parameter.

secretId :: Text

Specifies the secret with the version with the list of staging labels you want to modify. You can specify either the Amazon Resource Name (ARN) or the friendly name of the secret.

For an ARN, we recommend that you specify a complete ARN rather than a partial ARN.

versionStage :: Text

The staging label to add to this version.

newUpdateSecretVersionStage

Arguments

:: Text
$sel:secretId:UpdateSecretVersionStage'

-> Text
$sel:versionStage:UpdateSecretVersionStage'

-> UpdateSecretVersionStage

Create a value of UpdateSecretVersionStage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:removeFromVersionId:UpdateSecretVersionStage', updateSecretVersionStage_removeFromVersionId - Specifies the secret version ID of the version that the staging label is to be removed from. If the staging label you are trying to attach to one version is already attached to a different version, then you must include this parameter and specify the version that the label is to be removed from. If the label is attached and you either do not specify this parameter, or the version ID does not match, then the operation fails.

$sel:moveToVersionId:UpdateSecretVersionStage', updateSecretVersionStage_moveToVersionId - (Optional) The secret version ID that you want to add the staging label. If you want to remove a label from a version, then do not specify this parameter.

If the staging label is already attached to a different version of the secret, then you must also specify the RemoveFromVersionId parameter.

$sel:secretId:UpdateSecretVersionStage', updateSecretVersionStage_secretId - Specifies the secret with the version with the list of staging labels you want to modify. You can specify either the Amazon Resource Name (ARN) or the friendly name of the secret.

For an ARN, we recommend that you specify a complete ARN rather than a partial ARN.

$sel:versionStage:UpdateSecretVersionStage', updateSecretVersionStage_versionStage - The staging label to add to this version.

Request Lenses

updateSecretVersionStage_removeFromVersionId :: Lens' UpdateSecretVersionStage (Maybe Text)

Specifies the secret version ID of the version that the staging label is to be removed from. If the staging label you are trying to attach to one version is already attached to a different version, then you must include this parameter and specify the version that the label is to be removed from. If the label is attached and you either do not specify this parameter, or the version ID does not match, then the operation fails.

updateSecretVersionStage_moveToVersionId :: Lens' UpdateSecretVersionStage (Maybe Text)

(Optional) The secret version ID that you want to add the staging label. If you want to remove a label from a version, then do not specify this parameter.

If the staging label is already attached to a different version of the secret, then you must also specify the RemoveFromVersionId parameter.

updateSecretVersionStage_secretId :: Lens' UpdateSecretVersionStage Text

Specifies the secret with the version with the list of staging labels you want to modify. You can specify either the Amazon Resource Name (ARN) or the friendly name of the secret.

For an ARN, we recommend that you specify a complete ARN rather than a partial ARN.

updateSecretVersionStage_versionStage :: Lens' UpdateSecretVersionStage Text

The staging label to add to this version.

Destructuring the Response

data UpdateSecretVersionStageResponse

See: newUpdateSecretVersionStageResponse smart constructor.

Constructors

UpdateSecretVersionStageResponse'

Fields

arn :: Maybe Text

The ARN of the secret with the modified staging label.

name :: Maybe Text

The friendly name of the secret with the modified staging label.

httpStatus :: Int

The response's http status code.

newUpdateSecretVersionStageResponse

Arguments

:: Int
$sel:httpStatus:UpdateSecretVersionStageResponse'

-> UpdateSecretVersionStageResponse

Create a value of UpdateSecretVersionStageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:UpdateSecretVersionStageResponse', updateSecretVersionStageResponse_arn - The ARN of the secret with the modified staging label.

$sel:name:UpdateSecretVersionStageResponse', updateSecretVersionStageResponse_name - The friendly name of the secret with the modified staging label.

$sel:httpStatus:UpdateSecretVersionStageResponse', updateSecretVersionStageResponse_httpStatus - The response's http status code.

Response Lenses

updateSecretVersionStageResponse_arn :: Lens' UpdateSecretVersionStageResponse (Maybe Text)

The ARN of the secret with the modified staging label.

updateSecretVersionStageResponse_name :: Lens' UpdateSecretVersionStageResponse (Maybe Text)

The friendly name of the secret with the modified staging label.

updateSecretVersionStageResponse_httpStatus :: Lens' UpdateSecretVersionStageResponse Int

The response's http status code.