Network.AWS.SecurityHub.BatchDisableStandards

Description

Disables the standards specified by the provided StandardsSubscriptionArns.

For more information, see Security Standards section of the Security Hub User Guide.

Creating a Request

data BatchDisableStandards

See: newBatchDisableStandards smart constructor.

Constructors

BatchDisableStandards'

Fields

standardsSubscriptionArns :: NonEmpty Text

The ARNs of the standards subscriptions to disable.

newBatchDisableStandards

Arguments

:: NonEmpty Text
$sel:standardsSubscriptionArns:BatchDisableStandards'

-> BatchDisableStandards

Create a value of BatchDisableStandards with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:standardsSubscriptionArns:BatchDisableStandards', batchDisableStandards_standardsSubscriptionArns - The ARNs of the standards subscriptions to disable.

Request Lenses

batchDisableStandards_standardsSubscriptionArns :: Lens' BatchDisableStandards (NonEmpty Text)

The ARNs of the standards subscriptions to disable.

Destructuring the Response

data BatchDisableStandardsResponse

See: newBatchDisableStandardsResponse smart constructor.

Constructors

BatchDisableStandardsResponse'

Fields

standardsSubscriptions :: Maybe [StandardsSubscription]

The details of the standards subscriptions that were disabled.

httpStatus :: Int

The response's http status code.

newBatchDisableStandardsResponse

Arguments

:: Int
$sel:httpStatus:BatchDisableStandardsResponse'

-> BatchDisableStandardsResponse

Create a value of BatchDisableStandardsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:standardsSubscriptions:BatchDisableStandardsResponse', batchDisableStandardsResponse_standardsSubscriptions - The details of the standards subscriptions that were disabled.

$sel:httpStatus:BatchDisableStandardsResponse', batchDisableStandardsResponse_httpStatus - The response's http status code.

Response Lenses

batchDisableStandardsResponse_standardsSubscriptions :: Lens' BatchDisableStandardsResponse (Maybe [StandardsSubscription])

The details of the standards subscriptions that were disabled.

batchDisableStandardsResponse_httpStatus :: Lens' BatchDisableStandardsResponse Int

The response's http status code.