Network.AWS.SecurityHub.BatchUpdateFindings

Description

Used by Security Hub customers to update information about their investigation into a finding. Requested by administrator accounts or member accounts. Administrator accounts can update findings for their account and their member accounts. Member accounts can update findings for their account.

Updates from BatchUpdateFindings do not affect the value of UpdatedAt for a finding.

Administrator and member accounts can use BatchUpdateFindings to update the following finding fields and objects.

Confidence

Criticality

Note

RelatedFindings

Severity

Types

UserDefinedFields

VerificationState

Workflow

You can configure IAM policies to restrict access to fields and field values. For example, you might not want member accounts to be able to suppress findings or change the finding severity. See Configuring access to BatchUpdateFindings in the Security Hub User Guide.

Creating a Request

data BatchUpdateFindings

See: newBatchUpdateFindings smart constructor.

Constructors

BatchUpdateFindings'

Fields

criticality :: Maybe Natural

The updated value for the level of importance assigned to the resources associated with the findings.

A score of 0 means that the underlying resources have no criticality, and a score of 100 is reserved for the most critical resources.

note :: Maybe NoteUpdate

severity :: Maybe SeverityUpdate

Used to update the finding severity.

types :: Maybe [Text]

One or more finding types in the format of namespace/category/classifier that classify a finding.

Valid namespace values are as follows.

Software and Configuration Checks

TTPs

Effects

Unusual Behaviors

Sensitive Data Identifications

relatedFindings :: Maybe [RelatedFinding]

A list of findings that are related to the updated findings.

confidence :: Maybe Natural

The updated value for the finding confidence. Confidence is defined as the likelihood that a finding accurately identifies the behavior or issue that it was intended to identify.

Confidence is scored on a 0-100 basis using a ratio scale, where 0 means zero percent confidence and 100 means 100 percent confidence.

workflow :: Maybe WorkflowUpdate

Used to update the workflow status of a finding.

The workflow status indicates the progress of the investigation into the finding.

verificationState :: Maybe VerificationState

Indicates the veracity of a finding.

The available values for VerificationState are as follows.

UNKNOWN – The default disposition of a security finding

TRUE_POSITIVE – The security finding is confirmed

FALSE_POSITIVE – The security finding was determined to be a false alarm

BENIGN_POSITIVE – A special case of TRUE_POSITIVE where the finding doesn't pose any threat, is expected, or both

userDefinedFields :: Maybe (HashMap Text Text)

A list of name/value string pairs associated with the finding. These are custom, user-defined fields added to a finding.

findingIdentifiers :: [AwsSecurityFindingIdentifier]

The list of findings to update. BatchUpdateFindings can be used to update up to 100 findings at a time.

For each finding, the list provides the finding identifier and the ARN of the finding provider.

newBatchUpdateFindings :: BatchUpdateFindings

Create a value of BatchUpdateFindings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:criticality:BatchUpdateFindings', batchUpdateFindings_criticality - The updated value for the level of importance assigned to the resources associated with the findings.

A score of 0 means that the underlying resources have no criticality, and a score of 100 is reserved for the most critical resources.

$sel:note:BatchUpdateFindings', batchUpdateFindings_note - Undocumented member.

$sel:severity:BatchUpdateFindings', batchUpdateFindings_severity - Used to update the finding severity.

$sel:types:BatchUpdateFindings', batchUpdateFindings_types - One or more finding types in the format of namespace/category/classifier that classify a finding.

Valid namespace values are as follows.

Software and Configuration Checks

TTPs

Effects

Unusual Behaviors

Sensitive Data Identifications

$sel:relatedFindings:BatchUpdateFindings', batchUpdateFindings_relatedFindings - A list of findings that are related to the updated findings.

$sel:confidence:BatchUpdateFindings', batchUpdateFindings_confidence - The updated value for the finding confidence. Confidence is defined as the likelihood that a finding accurately identifies the behavior or issue that it was intended to identify.

Confidence is scored on a 0-100 basis using a ratio scale, where 0 means zero percent confidence and 100 means 100 percent confidence.

$sel:workflow:BatchUpdateFindings', batchUpdateFindings_workflow - Used to update the workflow status of a finding.

The workflow status indicates the progress of the investigation into the finding.

$sel:verificationState:BatchUpdateFindings', batchUpdateFindings_verificationState - Indicates the veracity of a finding.

The available values for VerificationState are as follows.

UNKNOWN – The default disposition of a security finding

TRUE_POSITIVE – The security finding is confirmed

FALSE_POSITIVE – The security finding was determined to be a false alarm

BENIGN_POSITIVE – A special case of TRUE_POSITIVE where the finding doesn't pose any threat, is expected, or both

$sel:userDefinedFields:BatchUpdateFindings', batchUpdateFindings_userDefinedFields - A list of name/value string pairs associated with the finding. These are custom, user-defined fields added to a finding.

$sel:findingIdentifiers:BatchUpdateFindings', batchUpdateFindings_findingIdentifiers - The list of findings to update. BatchUpdateFindings can be used to update up to 100 findings at a time.

For each finding, the list provides the finding identifier and the ARN of the finding provider.

Request Lenses

batchUpdateFindings_criticality :: Lens' BatchUpdateFindings (Maybe Natural)

The updated value for the level of importance assigned to the resources associated with the findings.

A score of 0 means that the underlying resources have no criticality, and a score of 100 is reserved for the most critical resources.

batchUpdateFindings_note :: Lens' BatchUpdateFindings (Maybe NoteUpdate)

Undocumented member.

batchUpdateFindings_severity :: Lens' BatchUpdateFindings (Maybe SeverityUpdate)

Used to update the finding severity.

batchUpdateFindings_types :: Lens' BatchUpdateFindings (Maybe [Text])

One or more finding types in the format of namespace/category/classifier that classify a finding.

Valid namespace values are as follows.

Software and Configuration Checks

TTPs

Effects

Unusual Behaviors

Sensitive Data Identifications

batchUpdateFindings_relatedFindings :: Lens' BatchUpdateFindings (Maybe [RelatedFinding])

A list of findings that are related to the updated findings.

batchUpdateFindings_confidence :: Lens' BatchUpdateFindings (Maybe Natural)

The updated value for the finding confidence. Confidence is defined as the likelihood that a finding accurately identifies the behavior or issue that it was intended to identify.

Confidence is scored on a 0-100 basis using a ratio scale, where 0 means zero percent confidence and 100 means 100 percent confidence.

batchUpdateFindings_workflow :: Lens' BatchUpdateFindings (Maybe WorkflowUpdate)

Used to update the workflow status of a finding.

The workflow status indicates the progress of the investigation into the finding.

batchUpdateFindings_verificationState :: Lens' BatchUpdateFindings (Maybe VerificationState)

Indicates the veracity of a finding.

The available values for VerificationState are as follows.

UNKNOWN – The default disposition of a security finding

TRUE_POSITIVE – The security finding is confirmed

FALSE_POSITIVE – The security finding was determined to be a false alarm

BENIGN_POSITIVE – A special case of TRUE_POSITIVE where the finding doesn't pose any threat, is expected, or both

batchUpdateFindings_userDefinedFields :: Lens' BatchUpdateFindings (Maybe (HashMap Text Text))

A list of name/value string pairs associated with the finding. These are custom, user-defined fields added to a finding.

batchUpdateFindings_findingIdentifiers :: Lens' BatchUpdateFindings [AwsSecurityFindingIdentifier]

The list of findings to update. BatchUpdateFindings can be used to update up to 100 findings at a time.

For each finding, the list provides the finding identifier and the ARN of the finding provider.

Destructuring the Response

data BatchUpdateFindingsResponse

See: newBatchUpdateFindingsResponse smart constructor.

Constructors

BatchUpdateFindingsResponse'

Fields

httpStatus :: Int

The response's http status code.

processedFindings :: [AwsSecurityFindingIdentifier]

The list of findings that were updated successfully.

unprocessedFindings :: [BatchUpdateFindingsUnprocessedFinding]

The list of findings that were not updated.

newBatchUpdateFindingsResponse

Arguments

:: Int
$sel:httpStatus:BatchUpdateFindingsResponse'

-> BatchUpdateFindingsResponse

Create a value of BatchUpdateFindingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:BatchUpdateFindingsResponse', batchUpdateFindingsResponse_httpStatus - The response's http status code.

$sel:processedFindings:BatchUpdateFindingsResponse', batchUpdateFindingsResponse_processedFindings - The list of findings that were updated successfully.

$sel:unprocessedFindings:BatchUpdateFindingsResponse', batchUpdateFindingsResponse_unprocessedFindings - The list of findings that were not updated.

Response Lenses

batchUpdateFindingsResponse_httpStatus :: Lens' BatchUpdateFindingsResponse Int

The response's http status code.

batchUpdateFindingsResponse_processedFindings :: Lens' BatchUpdateFindingsResponse [AwsSecurityFindingIdentifier]

The list of findings that were updated successfully.

batchUpdateFindingsResponse_unprocessedFindings :: Lens' BatchUpdateFindingsResponse [BatchUpdateFindingsUnprocessedFinding]

The list of findings that were not updated.