Network.AWS.SecurityHub.DeleteFindingAggregator

Description

Deletes a finding aggregator. When you delete the finding aggregator, you stop finding aggregation.

When you stop finding aggregation, findings that were already aggregated to the aggregation Region are still visible from the aggregation Region. New findings and finding updates are not aggregated.

Creating a Request

data DeleteFindingAggregator

See: newDeleteFindingAggregator smart constructor.

Constructors

DeleteFindingAggregator'

Fields

findingAggregatorArn :: Text

The ARN of the finding aggregator to delete. To obtain the ARN, use ListFindingAggregators.

newDeleteFindingAggregator

Arguments

:: Text
$sel:findingAggregatorArn:DeleteFindingAggregator'

-> DeleteFindingAggregator

Create a value of DeleteFindingAggregator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:findingAggregatorArn:DeleteFindingAggregator', deleteFindingAggregator_findingAggregatorArn - The ARN of the finding aggregator to delete. To obtain the ARN, use ListFindingAggregators.

Request Lenses

deleteFindingAggregator_findingAggregatorArn :: Lens' DeleteFindingAggregator Text

The ARN of the finding aggregator to delete. To obtain the ARN, use ListFindingAggregators.

Destructuring the Response

data DeleteFindingAggregatorResponse

See: newDeleteFindingAggregatorResponse smart constructor.

Constructors

DeleteFindingAggregatorResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteFindingAggregatorResponse

Arguments

:: Int
$sel:httpStatus:DeleteFindingAggregatorResponse'

-> DeleteFindingAggregatorResponse

Create a value of DeleteFindingAggregatorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteFindingAggregatorResponse', deleteFindingAggregatorResponse_httpStatus - The response's http status code.

Response Lenses

deleteFindingAggregatorResponse_httpStatus :: Lens' DeleteFindingAggregatorResponse Int

The response's http status code.