Network.AWS.SecurityHub.DeleteMembers

Description

Deletes the specified member accounts from Security Hub.

Can be used to delete member accounts that belong to an organization as well as member accounts that were invited manually.

Creating a Request

data DeleteMembers

See: newDeleteMembers smart constructor.

Constructors

DeleteMembers'

Fields

accountIds :: [Text]

The list of account IDs for the member accounts to delete.

newDeleteMembers :: DeleteMembers

Create a value of DeleteMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountIds:DeleteMembers', deleteMembers_accountIds - The list of account IDs for the member accounts to delete.

Request Lenses

deleteMembers_accountIds :: Lens' DeleteMembers [Text]

The list of account IDs for the member accounts to delete.

Destructuring the Response

data DeleteMembersResponse

See: newDeleteMembersResponse smart constructor.

Constructors

DeleteMembersResponse'

Fields

unprocessedAccounts :: Maybe [Result]

The list of Amazon Web Services accounts that were not deleted. For each account, the list includes the account ID and the email address.

httpStatus :: Int

The response's http status code.

newDeleteMembersResponse

Arguments

:: Int
$sel:httpStatus:DeleteMembersResponse'

-> DeleteMembersResponse

Create a value of DeleteMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unprocessedAccounts:DeleteMembersResponse', deleteMembersResponse_unprocessedAccounts - The list of Amazon Web Services accounts that were not deleted. For each account, the list includes the account ID and the email address.

$sel:httpStatus:DeleteMembersResponse', deleteMembersResponse_httpStatus - The response's http status code.

Response Lenses

deleteMembersResponse_unprocessedAccounts :: Lens' DeleteMembersResponse (Maybe [Result])

The list of Amazon Web Services accounts that were not deleted. For each account, the list includes the account ID and the email address.

deleteMembersResponse_httpStatus :: Lens' DeleteMembersResponse Int

The response's http status code.