Network.AWS.SecurityHub.EnableImportFindingsForProduct

Description

Enables the integration of a partner product with Security Hub. Integrated products send findings to Security Hub.

When you enable a product integration, a permissions policy that grants permission for the product to send findings to Security Hub is applied.

Creating a Request

data EnableImportFindingsForProduct

See: newEnableImportFindingsForProduct smart constructor.

Constructors

EnableImportFindingsForProduct'

Fields

productArn :: Text

The ARN of the product to enable the integration for.

newEnableImportFindingsForProduct

Arguments

:: Text
$sel:productArn:EnableImportFindingsForProduct'

-> EnableImportFindingsForProduct

Create a value of EnableImportFindingsForProduct with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:productArn:EnableImportFindingsForProduct', enableImportFindingsForProduct_productArn - The ARN of the product to enable the integration for.

Request Lenses

enableImportFindingsForProduct_productArn :: Lens' EnableImportFindingsForProduct Text

The ARN of the product to enable the integration for.

Destructuring the Response

data EnableImportFindingsForProductResponse

See: newEnableImportFindingsForProductResponse smart constructor.

Constructors

EnableImportFindingsForProductResponse'

Fields

productSubscriptionArn :: Maybe Text

The ARN of your subscription to the product to enable integrations for.

httpStatus :: Int

The response's http status code.

newEnableImportFindingsForProductResponse

Arguments

:: Int
$sel:httpStatus:EnableImportFindingsForProductResponse'

-> EnableImportFindingsForProductResponse

Create a value of EnableImportFindingsForProductResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:productSubscriptionArn:EnableImportFindingsForProductResponse', enableImportFindingsForProductResponse_productSubscriptionArn - The ARN of your subscription to the product to enable integrations for.

$sel:httpStatus:EnableImportFindingsForProductResponse', enableImportFindingsForProductResponse_httpStatus - The response's http status code.

Response Lenses

enableImportFindingsForProductResponse_productSubscriptionArn :: Lens' EnableImportFindingsForProductResponse (Maybe Text)

The ARN of your subscription to the product to enable integrations for.

enableImportFindingsForProductResponse_httpStatus :: Lens' EnableImportFindingsForProductResponse Int

The response's http status code.