Network.AWS.SecurityHub.GetEnabledStandards

Description

Returns a list of the standards that are currently enabled.

This operation returns paginated results.

Creating a Request

data GetEnabledStandards

See: newGetEnabledStandards smart constructor.

Constructors

GetEnabledStandards'

Fields

nextToken :: Maybe Text

The token that is required for pagination. On your first call to the GetEnabledStandards operation, set the value of this parameter to NULL.

For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

standardsSubscriptionArns :: Maybe (NonEmpty Text)

The list of the standards subscription ARNs for the standards to retrieve.

maxResults :: Maybe Natural

The maximum number of results to return in the response.

newGetEnabledStandards :: GetEnabledStandards

Create a value of GetEnabledStandards with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetEnabledStandards', getEnabledStandards_nextToken - The token that is required for pagination. On your first call to the GetEnabledStandards operation, set the value of this parameter to NULL.

For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

$sel:standardsSubscriptionArns:GetEnabledStandards', getEnabledStandards_standardsSubscriptionArns - The list of the standards subscription ARNs for the standards to retrieve.

$sel:maxResults:GetEnabledStandards', getEnabledStandards_maxResults - The maximum number of results to return in the response.

Request Lenses

getEnabledStandards_nextToken :: Lens' GetEnabledStandards (Maybe Text)

The token that is required for pagination. On your first call to the GetEnabledStandards operation, set the value of this parameter to NULL.

For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

getEnabledStandards_standardsSubscriptionArns :: Lens' GetEnabledStandards (Maybe (NonEmpty Text))

The list of the standards subscription ARNs for the standards to retrieve.

getEnabledStandards_maxResults :: Lens' GetEnabledStandards (Maybe Natural)

The maximum number of results to return in the response.

Destructuring the Response

data GetEnabledStandardsResponse

See: newGetEnabledStandardsResponse smart constructor.

Constructors

GetEnabledStandardsResponse'

Fields

nextToken :: Maybe Text

The pagination token to use to request the next page of results.

standardsSubscriptions :: Maybe [StandardsSubscription]

The list of StandardsSubscriptions objects that include information about the enabled standards.

httpStatus :: Int

The response's http status code.

newGetEnabledStandardsResponse

Arguments

:: Int
$sel:httpStatus:GetEnabledStandardsResponse'

-> GetEnabledStandardsResponse

Create a value of GetEnabledStandardsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetEnabledStandards', getEnabledStandardsResponse_nextToken - The pagination token to use to request the next page of results.

$sel:standardsSubscriptions:GetEnabledStandardsResponse', getEnabledStandardsResponse_standardsSubscriptions - The list of StandardsSubscriptions objects that include information about the enabled standards.

$sel:httpStatus:GetEnabledStandardsResponse', getEnabledStandardsResponse_httpStatus - The response's http status code.

Response Lenses

getEnabledStandardsResponse_nextToken :: Lens' GetEnabledStandardsResponse (Maybe Text)

The pagination token to use to request the next page of results.

getEnabledStandardsResponse_standardsSubscriptions :: Lens' GetEnabledStandardsResponse (Maybe [StandardsSubscription])

The list of StandardsSubscriptions objects that include information about the enabled standards.

getEnabledStandardsResponse_httpStatus :: Lens' GetEnabledStandardsResponse Int

The response's http status code.