Network.AWS.SecurityHub.GetInsights

Description

Lists and describes insights for the specified insight ARNs.

This operation returns paginated results.

Creating a Request

data GetInsights

See: newGetInsights smart constructor.

Constructors

GetInsights'

Fields

nextToken :: Maybe Text

The token that is required for pagination. On your first call to the GetInsights operation, set the value of this parameter to NULL.

For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

insightArns :: Maybe [Text]

The ARNs of the insights to describe. If you do not provide any insight ARNs, then GetInsights returns all of your custom insights. It does not return any managed insights.

maxResults :: Maybe Natural

The maximum number of items to return in the response.

newGetInsights :: GetInsights

Create a value of GetInsights with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetInsights', getInsights_nextToken - The token that is required for pagination. On your first call to the GetInsights operation, set the value of this parameter to NULL.

For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

$sel:insightArns:GetInsights', getInsights_insightArns - The ARNs of the insights to describe. If you do not provide any insight ARNs, then GetInsights returns all of your custom insights. It does not return any managed insights.

$sel:maxResults:GetInsights', getInsights_maxResults - The maximum number of items to return in the response.

Request Lenses

getInsights_nextToken :: Lens' GetInsights (Maybe Text)

The token that is required for pagination. On your first call to the GetInsights operation, set the value of this parameter to NULL.

For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

getInsights_insightArns :: Lens' GetInsights (Maybe [Text])

The ARNs of the insights to describe. If you do not provide any insight ARNs, then GetInsights returns all of your custom insights. It does not return any managed insights.

getInsights_maxResults :: Lens' GetInsights (Maybe Natural)

The maximum number of items to return in the response.

Destructuring the Response

data GetInsightsResponse

See: newGetInsightsResponse smart constructor.

Constructors

GetInsightsResponse'

Fields

nextToken :: Maybe Text

The pagination token to use to request the next page of results.

httpStatus :: Int

The response's http status code.

insights :: [Insight]

The insights returned by the operation.

newGetInsightsResponse

Arguments

:: Int
$sel:httpStatus:GetInsightsResponse'

-> GetInsightsResponse

Create a value of GetInsightsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetInsights', getInsightsResponse_nextToken - The pagination token to use to request the next page of results.

$sel:httpStatus:GetInsightsResponse', getInsightsResponse_httpStatus - The response's http status code.

$sel:insights:GetInsightsResponse', getInsightsResponse_insights - The insights returned by the operation.

Response Lenses

getInsightsResponse_nextToken :: Lens' GetInsightsResponse (Maybe Text)

The pagination token to use to request the next page of results.

getInsightsResponse_httpStatus :: Lens' GetInsightsResponse Int

The response's http status code.

getInsightsResponse_insights :: Lens' GetInsightsResponse [Insight]

The insights returned by the operation.