Network.AWS.SecurityHub.GetMembers

Description

Returns the details for the Security Hub member accounts for the specified account IDs.

An administrator account can be either the delegated Security Hub administrator account for an organization or an administrator account that enabled Security Hub manually.

The results include both member accounts that are managed using Organizations and accounts that were invited manually.

Creating a Request

data GetMembers

See: newGetMembers smart constructor.

Constructors

GetMembers'

Fields

accountIds :: [Text]

The list of account IDs for the Security Hub member accounts to return the details for.

newGetMembers :: GetMembers

Create a value of GetMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountIds:GetMembers', getMembers_accountIds - The list of account IDs for the Security Hub member accounts to return the details for.

Request Lenses

getMembers_accountIds :: Lens' GetMembers [Text]

The list of account IDs for the Security Hub member accounts to return the details for.

Destructuring the Response

data GetMembersResponse

See: newGetMembersResponse smart constructor.

Constructors

GetMembersResponse'

Fields

members :: Maybe [Member]

The list of details about the Security Hub member accounts.

unprocessedAccounts :: Maybe [Result]

The list of Amazon Web Services accounts that could not be processed. For each account, the list includes the account ID and the email address.

httpStatus :: Int

The response's http status code.

newGetMembersResponse

Arguments

:: Int
$sel:httpStatus:GetMembersResponse'

-> GetMembersResponse

Create a value of GetMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:members:GetMembersResponse', getMembersResponse_members - The list of details about the Security Hub member accounts.

$sel:unprocessedAccounts:GetMembersResponse', getMembersResponse_unprocessedAccounts - The list of Amazon Web Services accounts that could not be processed. For each account, the list includes the account ID and the email address.

$sel:httpStatus:GetMembersResponse', getMembersResponse_httpStatus - The response's http status code.

Response Lenses

getMembersResponse_members :: Lens' GetMembersResponse (Maybe [Member])

The list of details about the Security Hub member accounts.

getMembersResponse_unprocessedAccounts :: Lens' GetMembersResponse (Maybe [Result])

The list of Amazon Web Services accounts that could not be processed. For each account, the list includes the account ID and the email address.

getMembersResponse_httpStatus :: Lens' GetMembersResponse Int

The response's http status code.