Network.AWS.SecurityHub.InviteMembers

Description

Invites other Amazon Web Services accounts to become member accounts for the Security Hub administrator account that the invitation is sent from.

This operation is only used to invite accounts that do not belong to an organization. Organization accounts do not receive invitations.

Before you can use this action to invite a member, you must first use the CreateMembers action to create the member account in Security Hub.

When the account owner enables Security Hub and accepts the invitation to become a member account, the administrator account can view the findings generated from the member account.

Creating a Request

data InviteMembers

See: newInviteMembers smart constructor.

Constructors

InviteMembers'

Fields

accountIds :: [Text]

The list of account IDs of the Amazon Web Services accounts to invite to Security Hub as members.

newInviteMembers :: InviteMembers

Create a value of InviteMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountIds:InviteMembers', inviteMembers_accountIds - The list of account IDs of the Amazon Web Services accounts to invite to Security Hub as members.

Request Lenses

inviteMembers_accountIds :: Lens' InviteMembers [Text]

The list of account IDs of the Amazon Web Services accounts to invite to Security Hub as members.

Destructuring the Response

data InviteMembersResponse

See: newInviteMembersResponse smart constructor.

Constructors

InviteMembersResponse'

Fields

unprocessedAccounts :: Maybe [Result]

The list of Amazon Web Services accounts that could not be processed. For each account, the list includes the account ID and the email address.

httpStatus :: Int

The response's http status code.

newInviteMembersResponse

Arguments

:: Int
$sel:httpStatus:InviteMembersResponse'

-> InviteMembersResponse

Create a value of InviteMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unprocessedAccounts:InviteMembersResponse', inviteMembersResponse_unprocessedAccounts - The list of Amazon Web Services accounts that could not be processed. For each account, the list includes the account ID and the email address.

$sel:httpStatus:InviteMembersResponse', inviteMembersResponse_httpStatus - The response's http status code.

Response Lenses

inviteMembersResponse_unprocessedAccounts :: Lens' InviteMembersResponse (Maybe [Result])

The list of Amazon Web Services accounts that could not be processed. For each account, the list includes the account ID and the email address.

inviteMembersResponse_httpStatus :: Lens' InviteMembersResponse Int

The response's http status code.