Network.AWS.SecurityHub.Types.ActionRemoteIpDetails

Description

Documentation

data ActionRemoteIpDetails

For AwsApiAction, NetworkConnectionAction, and PortProbeAction, RemoteIpDetails provides information about the remote IP address that was involved in the action.

See: newActionRemoteIpDetails smart constructor.

Constructors

ActionRemoteIpDetails'

Fields

country :: Maybe Country

The country where the remote IP address is located.

city :: Maybe City

The city where the remote IP address is located.

ipAddressV4 :: Maybe Text

The IP address.

geoLocation :: Maybe GeoLocation

The coordinates of the location of the remote IP address.

organization :: Maybe IpOrganizationDetails

The internet service provider (ISP) organization associated with the remote IP address.

newActionRemoteIpDetails :: ActionRemoteIpDetails

Create a value of ActionRemoteIpDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:country:ActionRemoteIpDetails', actionRemoteIpDetails_country - The country where the remote IP address is located.

$sel:city:ActionRemoteIpDetails', actionRemoteIpDetails_city - The city where the remote IP address is located.

$sel:ipAddressV4:ActionRemoteIpDetails', actionRemoteIpDetails_ipAddressV4 - The IP address.

$sel:geoLocation:ActionRemoteIpDetails', actionRemoteIpDetails_geoLocation - The coordinates of the location of the remote IP address.

$sel:organization:ActionRemoteIpDetails', actionRemoteIpDetails_organization - The internet service provider (ISP) organization associated with the remote IP address.

actionRemoteIpDetails_country :: Lens' ActionRemoteIpDetails (Maybe Country)

The country where the remote IP address is located.

actionRemoteIpDetails_city :: Lens' ActionRemoteIpDetails (Maybe City)

The city where the remote IP address is located.

actionRemoteIpDetails_ipAddressV4 :: Lens' ActionRemoteIpDetails (Maybe Text)

The IP address.

actionRemoteIpDetails_geoLocation :: Lens' ActionRemoteIpDetails (Maybe GeoLocation)

The coordinates of the location of the remote IP address.

actionRemoteIpDetails_organization :: Lens' ActionRemoteIpDetails (Maybe IpOrganizationDetails)

The internet service provider (ISP) organization associated with the remote IP address.