Network.AWS.SecurityHub.Types.AwsApiGatewayMethodSettings

Description

Documentation

data AwsApiGatewayMethodSettings

Defines settings for a method for the stage.

See: newAwsApiGatewayMethodSettings smart constructor.

Constructors

AwsApiGatewayMethodSettings'

Fields

dataTraceEnabled :: Maybe Bool

Indicates whether data trace logging is enabled for the method. Data trace logging affects the log entries that are pushed to CloudWatch Logs.

cacheTtlInSeconds :: Maybe Int

Specifies the time to live (TTL), in seconds, for cached responses. The higher the TTL, the longer the response is cached.

httpMethod :: Maybe Text

The HTTP method. You can use an asterisk (*) as a wildcard to apply method settings to multiple methods.

throttlingBurstLimit :: Maybe Int

The throttling burst limit for the method.

cacheDataEncrypted :: Maybe Bool

Indicates whether the cached responses are encrypted.

loggingLevel :: Maybe Text

The logging level for this method. The logging level affects the log entries that are pushed to CloudWatch Logs.

If the logging level is ERROR, then the logs only include error-level entries.

If the logging level is INFO, then the logs include both ERROR events and extra informational events.

Valid values: OFF | ERROR | INFO

requireAuthorizationForCacheControl :: Maybe Bool

Indicates whether authorization is required for a cache invalidation request.

cachingEnabled :: Maybe Bool

Indicates whether responses are cached and returned for requests. For responses to be cached, a cache cluster must be enabled on the stage.

resourcePath :: Maybe Text

The resource path for this method. Forward slashes (/) are encoded as ~1 . The initial slash must include a forward slash.

For example, the path value /resource/subresource must be encoded as /~1resource~1subresource.

To specify the root path, use only a slash (/). You can use an asterisk (*) as a wildcard to apply method settings to multiple methods.

throttlingRateLimit :: Maybe Double

The throttling rate limit for the method.

metricsEnabled :: Maybe Bool

Indicates whether CloudWatch metrics are enabled for the method.

unauthorizedCacheControlHeaderStrategy :: Maybe Text

Indicates how to handle unauthorized requests for cache invalidation.

Valid values: FAIL_WITH_403 | SUCCEED_WITH_RESPONSE_HEADER | SUCCEED_WITHOUT_RESPONSE_HEADER

newAwsApiGatewayMethodSettings :: AwsApiGatewayMethodSettings

Create a value of AwsApiGatewayMethodSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataTraceEnabled:AwsApiGatewayMethodSettings', awsApiGatewayMethodSettings_dataTraceEnabled - Indicates whether data trace logging is enabled for the method. Data trace logging affects the log entries that are pushed to CloudWatch Logs.

$sel:cacheTtlInSeconds:AwsApiGatewayMethodSettings', awsApiGatewayMethodSettings_cacheTtlInSeconds - Specifies the time to live (TTL), in seconds, for cached responses. The higher the TTL, the longer the response is cached.

$sel:httpMethod:AwsApiGatewayMethodSettings', awsApiGatewayMethodSettings_httpMethod - The HTTP method. You can use an asterisk (*) as a wildcard to apply method settings to multiple methods.

$sel:throttlingBurstLimit:AwsApiGatewayMethodSettings', awsApiGatewayMethodSettings_throttlingBurstLimit - The throttling burst limit for the method.

$sel:cacheDataEncrypted:AwsApiGatewayMethodSettings', awsApiGatewayMethodSettings_cacheDataEncrypted - Indicates whether the cached responses are encrypted.

$sel:loggingLevel:AwsApiGatewayMethodSettings', awsApiGatewayMethodSettings_loggingLevel - The logging level for this method. The logging level affects the log entries that are pushed to CloudWatch Logs.

If the logging level is ERROR, then the logs only include error-level entries.

If the logging level is INFO, then the logs include both ERROR events and extra informational events.

Valid values: OFF | ERROR | INFO

$sel:requireAuthorizationForCacheControl:AwsApiGatewayMethodSettings', awsApiGatewayMethodSettings_requireAuthorizationForCacheControl - Indicates whether authorization is required for a cache invalidation request.

$sel:cachingEnabled:AwsApiGatewayMethodSettings', awsApiGatewayMethodSettings_cachingEnabled - Indicates whether responses are cached and returned for requests. For responses to be cached, a cache cluster must be enabled on the stage.

$sel:resourcePath:AwsApiGatewayMethodSettings', awsApiGatewayMethodSettings_resourcePath - The resource path for this method. Forward slashes (/) are encoded as ~1 . The initial slash must include a forward slash.

For example, the path value /resource/subresource must be encoded as /~1resource~1subresource.

To specify the root path, use only a slash (/). You can use an asterisk (*) as a wildcard to apply method settings to multiple methods.

$sel:throttlingRateLimit:AwsApiGatewayMethodSettings', awsApiGatewayMethodSettings_throttlingRateLimit - The throttling rate limit for the method.

$sel:metricsEnabled:AwsApiGatewayMethodSettings', awsApiGatewayMethodSettings_metricsEnabled - Indicates whether CloudWatch metrics are enabled for the method.

$sel:unauthorizedCacheControlHeaderStrategy:AwsApiGatewayMethodSettings', awsApiGatewayMethodSettings_unauthorizedCacheControlHeaderStrategy - Indicates how to handle unauthorized requests for cache invalidation.

Valid values: FAIL_WITH_403 | SUCCEED_WITH_RESPONSE_HEADER | SUCCEED_WITHOUT_RESPONSE_HEADER

awsApiGatewayMethodSettings_dataTraceEnabled :: Lens' AwsApiGatewayMethodSettings (Maybe Bool)

Indicates whether data trace logging is enabled for the method. Data trace logging affects the log entries that are pushed to CloudWatch Logs.

awsApiGatewayMethodSettings_cacheTtlInSeconds :: Lens' AwsApiGatewayMethodSettings (Maybe Int)

Specifies the time to live (TTL), in seconds, for cached responses. The higher the TTL, the longer the response is cached.

awsApiGatewayMethodSettings_httpMethod :: Lens' AwsApiGatewayMethodSettings (Maybe Text)

The HTTP method. You can use an asterisk (*) as a wildcard to apply method settings to multiple methods.

awsApiGatewayMethodSettings_throttlingBurstLimit :: Lens' AwsApiGatewayMethodSettings (Maybe Int)

The throttling burst limit for the method.

awsApiGatewayMethodSettings_cacheDataEncrypted :: Lens' AwsApiGatewayMethodSettings (Maybe Bool)

Indicates whether the cached responses are encrypted.

awsApiGatewayMethodSettings_loggingLevel :: Lens' AwsApiGatewayMethodSettings (Maybe Text)

The logging level for this method. The logging level affects the log entries that are pushed to CloudWatch Logs.

If the logging level is ERROR, then the logs only include error-level entries.

If the logging level is INFO, then the logs include both ERROR events and extra informational events.

Valid values: OFF | ERROR | INFO

awsApiGatewayMethodSettings_requireAuthorizationForCacheControl :: Lens' AwsApiGatewayMethodSettings (Maybe Bool)

Indicates whether authorization is required for a cache invalidation request.

awsApiGatewayMethodSettings_cachingEnabled :: Lens' AwsApiGatewayMethodSettings (Maybe Bool)

Indicates whether responses are cached and returned for requests. For responses to be cached, a cache cluster must be enabled on the stage.

awsApiGatewayMethodSettings_resourcePath :: Lens' AwsApiGatewayMethodSettings (Maybe Text)

The resource path for this method. Forward slashes (/) are encoded as ~1 . The initial slash must include a forward slash.

For example, the path value /resource/subresource must be encoded as /~1resource~1subresource.

To specify the root path, use only a slash (/). You can use an asterisk (*) as a wildcard to apply method settings to multiple methods.

awsApiGatewayMethodSettings_throttlingRateLimit :: Lens' AwsApiGatewayMethodSettings (Maybe Double)

The throttling rate limit for the method.

awsApiGatewayMethodSettings_metricsEnabled :: Lens' AwsApiGatewayMethodSettings (Maybe Bool)

Indicates whether CloudWatch metrics are enabled for the method.

awsApiGatewayMethodSettings_unauthorizedCacheControlHeaderStrategy :: Lens' AwsApiGatewayMethodSettings (Maybe Text)

Indicates how to handle unauthorized requests for cache invalidation.

Valid values: FAIL_WITH_403 | SUCCEED_WITH_RESPONSE_HEADER | SUCCEED_WITHOUT_RESPONSE_HEADER