Network.AWS.SecurityHub.Types.AwsApiGatewayStageDetails

Description

Documentation

data AwsApiGatewayStageDetails

Provides information about a version 1 Amazon API Gateway stage.

See: newAwsApiGatewayStageDetails smart constructor.

Constructors

AwsApiGatewayStageDetails'

Fields

deploymentId :: Maybe Text

The identifier of the deployment that the stage points to.

accessLogSettings :: Maybe AwsApiGatewayAccessLogSettings

Settings for logging access for the stage.

variables :: Maybe (HashMap Text Text)

A map that defines the stage variables for the stage.

Variable names can have alphanumeric and underscore characters.

Variable values can contain the following characters:

Uppercase and lowercase letters

Numbers

Special characters -._~:/?#&=,

documentationVersion :: Maybe Text

The version of the API documentation that is associated with the stage.

clientCertificateId :: Maybe Text

The identifier of the client certificate for the stage.

tracingEnabled :: Maybe Bool

Indicates whether active tracing with X-Ray is enabled for the stage.

createdDate :: Maybe Text

Indicates when the stage was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

methodSettings :: Maybe [AwsApiGatewayMethodSettings]

Defines the method settings for the stage.

cacheClusterStatus :: Maybe Text

If a cache cluster is enabled, the status of the cache cluster.

webAclArn :: Maybe Text

The ARN of the web ACL associated with the stage.

stageName :: Maybe Text

The name of the stage.

cacheClusterEnabled :: Maybe Bool

Indicates whether a cache cluster is enabled for the stage.

cacheClusterSize :: Maybe Text

If a cache cluster is enabled, the size of the cache cluster.

lastUpdatedDate :: Maybe Text

Indicates when the stage was most recently updated.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

canarySettings :: Maybe AwsApiGatewayCanarySettings

Information about settings for canary deployment in the stage.

description :: Maybe Text

A description of the stage.

newAwsApiGatewayStageDetails :: AwsApiGatewayStageDetails

Create a value of AwsApiGatewayStageDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentId:AwsApiGatewayStageDetails', awsApiGatewayStageDetails_deploymentId - The identifier of the deployment that the stage points to.

$sel:accessLogSettings:AwsApiGatewayStageDetails', awsApiGatewayStageDetails_accessLogSettings - Settings for logging access for the stage.

$sel:variables:AwsApiGatewayStageDetails', awsApiGatewayStageDetails_variables - A map that defines the stage variables for the stage.

Variable names can have alphanumeric and underscore characters.

Variable values can contain the following characters:

Uppercase and lowercase letters

Numbers

Special characters -._~:/?#&=,

$sel:documentationVersion:AwsApiGatewayStageDetails', awsApiGatewayStageDetails_documentationVersion - The version of the API documentation that is associated with the stage.

$sel:clientCertificateId:AwsApiGatewayStageDetails', awsApiGatewayStageDetails_clientCertificateId - The identifier of the client certificate for the stage.

$sel:tracingEnabled:AwsApiGatewayStageDetails', awsApiGatewayStageDetails_tracingEnabled - Indicates whether active tracing with X-Ray is enabled for the stage.

$sel:createdDate:AwsApiGatewayStageDetails', awsApiGatewayStageDetails_createdDate - Indicates when the stage was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:methodSettings:AwsApiGatewayStageDetails', awsApiGatewayStageDetails_methodSettings - Defines the method settings for the stage.

$sel:cacheClusterStatus:AwsApiGatewayStageDetails', awsApiGatewayStageDetails_cacheClusterStatus - If a cache cluster is enabled, the status of the cache cluster.

$sel:webAclArn:AwsApiGatewayStageDetails', awsApiGatewayStageDetails_webAclArn - The ARN of the web ACL associated with the stage.

$sel:stageName:AwsApiGatewayStageDetails', awsApiGatewayStageDetails_stageName - The name of the stage.

$sel:cacheClusterEnabled:AwsApiGatewayStageDetails', awsApiGatewayStageDetails_cacheClusterEnabled - Indicates whether a cache cluster is enabled for the stage.

$sel:cacheClusterSize:AwsApiGatewayStageDetails', awsApiGatewayStageDetails_cacheClusterSize - If a cache cluster is enabled, the size of the cache cluster.

$sel:lastUpdatedDate:AwsApiGatewayStageDetails', awsApiGatewayStageDetails_lastUpdatedDate - Indicates when the stage was most recently updated.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:canarySettings:AwsApiGatewayStageDetails', awsApiGatewayStageDetails_canarySettings - Information about settings for canary deployment in the stage.

$sel:description:AwsApiGatewayStageDetails', awsApiGatewayStageDetails_description - A description of the stage.

awsApiGatewayStageDetails_deploymentId :: Lens' AwsApiGatewayStageDetails (Maybe Text)

The identifier of the deployment that the stage points to.

awsApiGatewayStageDetails_accessLogSettings :: Lens' AwsApiGatewayStageDetails (Maybe AwsApiGatewayAccessLogSettings)

Settings for logging access for the stage.

awsApiGatewayStageDetails_variables :: Lens' AwsApiGatewayStageDetails (Maybe (HashMap Text Text))

A map that defines the stage variables for the stage.

Variable names can have alphanumeric and underscore characters.

Variable values can contain the following characters:

Uppercase and lowercase letters

Numbers

Special characters -._~:/?#&=,

awsApiGatewayStageDetails_documentationVersion :: Lens' AwsApiGatewayStageDetails (Maybe Text)

The version of the API documentation that is associated with the stage.

awsApiGatewayStageDetails_clientCertificateId :: Lens' AwsApiGatewayStageDetails (Maybe Text)

The identifier of the client certificate for the stage.

awsApiGatewayStageDetails_tracingEnabled :: Lens' AwsApiGatewayStageDetails (Maybe Bool)

Indicates whether active tracing with X-Ray is enabled for the stage.

awsApiGatewayStageDetails_createdDate :: Lens' AwsApiGatewayStageDetails (Maybe Text)

Indicates when the stage was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsApiGatewayStageDetails_methodSettings :: Lens' AwsApiGatewayStageDetails (Maybe [AwsApiGatewayMethodSettings])

Defines the method settings for the stage.

awsApiGatewayStageDetails_cacheClusterStatus :: Lens' AwsApiGatewayStageDetails (Maybe Text)

If a cache cluster is enabled, the status of the cache cluster.

awsApiGatewayStageDetails_webAclArn :: Lens' AwsApiGatewayStageDetails (Maybe Text)

The ARN of the web ACL associated with the stage.

awsApiGatewayStageDetails_stageName :: Lens' AwsApiGatewayStageDetails (Maybe Text)

The name of the stage.

awsApiGatewayStageDetails_cacheClusterEnabled :: Lens' AwsApiGatewayStageDetails (Maybe Bool)

Indicates whether a cache cluster is enabled for the stage.

awsApiGatewayStageDetails_cacheClusterSize :: Lens' AwsApiGatewayStageDetails (Maybe Text)

If a cache cluster is enabled, the size of the cache cluster.

awsApiGatewayStageDetails_lastUpdatedDate :: Lens' AwsApiGatewayStageDetails (Maybe Text)

Indicates when the stage was most recently updated.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsApiGatewayStageDetails_canarySettings :: Lens' AwsApiGatewayStageDetails (Maybe AwsApiGatewayCanarySettings)

Information about settings for canary deployment in the stage.

awsApiGatewayStageDetails_description :: Lens' AwsApiGatewayStageDetails (Maybe Text)

A description of the stage.