Network.AWS.SecurityHub.Types.AwsApiGatewayV2StageDetails

Description

Documentation

data AwsApiGatewayV2StageDetails

Contains information about a version 2 stage for Amazon API Gateway.

See: newAwsApiGatewayV2StageDetails smart constructor.

Constructors

AwsApiGatewayV2StageDetails'

Fields

lastDeploymentStatusMessage :: Maybe Text

The status of the last deployment of a stage. Supported only if the stage has automatic deployment enabled.

deploymentId :: Maybe Text

The identifier of the deployment that the stage is associated with.

routeSettings :: Maybe AwsApiGatewayV2RouteSettings

The route settings for the stage.

accessLogSettings :: Maybe AwsApiGatewayAccessLogSettings

Information about settings for logging access for the stage.

clientCertificateId :: Maybe Text

The identifier of a client certificate for a stage. Supported only for WebSocket API calls.

stageVariables :: Maybe (HashMap Text Text)

A map that defines the stage variables for the stage.

Variable names can have alphanumeric and underscore characters.

Variable values can contain the following characters:

Uppercase and lowercase letters

Numbers

Special characters -._~:/?#&=,

autoDeploy :: Maybe Bool

Indicates whether updates to an API automatically trigger a new deployment.

createdDate :: Maybe Text

Indicates when the stage was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

defaultRouteSettings :: Maybe AwsApiGatewayV2RouteSettings

Default route settings for the stage.

apiGatewayManaged :: Maybe Bool

Indicates whether the stage is managed by API Gateway.

stageName :: Maybe Text

The name of the stage.

lastUpdatedDate :: Maybe Text

Indicates when the stage was most recently updated.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

description :: Maybe Text

The description of the stage.

newAwsApiGatewayV2StageDetails :: AwsApiGatewayV2StageDetails

Create a value of AwsApiGatewayV2StageDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastDeploymentStatusMessage:AwsApiGatewayV2StageDetails', awsApiGatewayV2StageDetails_lastDeploymentStatusMessage - The status of the last deployment of a stage. Supported only if the stage has automatic deployment enabled.

$sel:deploymentId:AwsApiGatewayV2StageDetails', awsApiGatewayV2StageDetails_deploymentId - The identifier of the deployment that the stage is associated with.

$sel:routeSettings:AwsApiGatewayV2StageDetails', awsApiGatewayV2StageDetails_routeSettings - The route settings for the stage.

$sel:accessLogSettings:AwsApiGatewayV2StageDetails', awsApiGatewayV2StageDetails_accessLogSettings - Information about settings for logging access for the stage.

$sel:clientCertificateId:AwsApiGatewayV2StageDetails', awsApiGatewayV2StageDetails_clientCertificateId - The identifier of a client certificate for a stage. Supported only for WebSocket API calls.

$sel:stageVariables:AwsApiGatewayV2StageDetails', awsApiGatewayV2StageDetails_stageVariables - A map that defines the stage variables for the stage.

Variable names can have alphanumeric and underscore characters.

Variable values can contain the following characters:

Uppercase and lowercase letters

Numbers

Special characters -._~:/?#&=,

$sel:autoDeploy:AwsApiGatewayV2StageDetails', awsApiGatewayV2StageDetails_autoDeploy - Indicates whether updates to an API automatically trigger a new deployment.

$sel:createdDate:AwsApiGatewayV2StageDetails', awsApiGatewayV2StageDetails_createdDate - Indicates when the stage was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:defaultRouteSettings:AwsApiGatewayV2StageDetails', awsApiGatewayV2StageDetails_defaultRouteSettings - Default route settings for the stage.

$sel:apiGatewayManaged:AwsApiGatewayV2StageDetails', awsApiGatewayV2StageDetails_apiGatewayManaged - Indicates whether the stage is managed by API Gateway.

$sel:stageName:AwsApiGatewayV2StageDetails', awsApiGatewayV2StageDetails_stageName - The name of the stage.

$sel:lastUpdatedDate:AwsApiGatewayV2StageDetails', awsApiGatewayV2StageDetails_lastUpdatedDate - Indicates when the stage was most recently updated.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:description:AwsApiGatewayV2StageDetails', awsApiGatewayV2StageDetails_description - The description of the stage.

awsApiGatewayV2StageDetails_lastDeploymentStatusMessage :: Lens' AwsApiGatewayV2StageDetails (Maybe Text)

The status of the last deployment of a stage. Supported only if the stage has automatic deployment enabled.

awsApiGatewayV2StageDetails_deploymentId :: Lens' AwsApiGatewayV2StageDetails (Maybe Text)

The identifier of the deployment that the stage is associated with.

awsApiGatewayV2StageDetails_routeSettings :: Lens' AwsApiGatewayV2StageDetails (Maybe AwsApiGatewayV2RouteSettings)

The route settings for the stage.

awsApiGatewayV2StageDetails_accessLogSettings :: Lens' AwsApiGatewayV2StageDetails (Maybe AwsApiGatewayAccessLogSettings)

Information about settings for logging access for the stage.

awsApiGatewayV2StageDetails_clientCertificateId :: Lens' AwsApiGatewayV2StageDetails (Maybe Text)

The identifier of a client certificate for a stage. Supported only for WebSocket API calls.

awsApiGatewayV2StageDetails_stageVariables :: Lens' AwsApiGatewayV2StageDetails (Maybe (HashMap Text Text))

A map that defines the stage variables for the stage.

Variable names can have alphanumeric and underscore characters.

Variable values can contain the following characters:

Uppercase and lowercase letters

Numbers

Special characters -._~:/?#&=,

awsApiGatewayV2StageDetails_autoDeploy :: Lens' AwsApiGatewayV2StageDetails (Maybe Bool)

Indicates whether updates to an API automatically trigger a new deployment.

awsApiGatewayV2StageDetails_createdDate :: Lens' AwsApiGatewayV2StageDetails (Maybe Text)

Indicates when the stage was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsApiGatewayV2StageDetails_defaultRouteSettings :: Lens' AwsApiGatewayV2StageDetails (Maybe AwsApiGatewayV2RouteSettings)

Default route settings for the stage.

awsApiGatewayV2StageDetails_apiGatewayManaged :: Lens' AwsApiGatewayV2StageDetails (Maybe Bool)

Indicates whether the stage is managed by API Gateway.

awsApiGatewayV2StageDetails_stageName :: Lens' AwsApiGatewayV2StageDetails (Maybe Text)

The name of the stage.

awsApiGatewayV2StageDetails_lastUpdatedDate :: Lens' AwsApiGatewayV2StageDetails (Maybe Text)

Indicates when the stage was most recently updated.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsApiGatewayV2StageDetails_description :: Lens' AwsApiGatewayV2StageDetails (Maybe Text)

The description of the stage.