Network.AWS.SecurityHub.Types.AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails

Description

Documentation

data AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails

Parameters that are used to automatically set up EBS volumes when an instance is launched.

See: newAwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails smart constructor.

Constructors

AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails'

Fields

deleteOnTermination :: Maybe Bool

Whether to delete the volume when the instance is terminated.

volumeSize :: Maybe Int

The volume size, in GiBs. The following are the supported volumes sizes for each volume type:

gp2 and gp3: 1-16,384

io1: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024

You must specify either SnapshotId or VolumeSize. If you specify both SnapshotId and VolumeSize, the volume size must be equal or greater than the size of the snapshot.

iops :: Maybe Int

The number of input/output (I/O) operations per second (IOPS) to provision for the volume.

Only supported for gp3 or io1 volumes. Required for io1 volumes. Not used with standard, gp2, st1, or sc1 volumes.

encrypted :: Maybe Bool

Whether to encrypt the volume.

volumeType :: Maybe Text

The volume type.

snapshotId :: Maybe Text

The snapshot ID of the volume to use.

You must specify either VolumeSize or SnapshotId.

newAwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails :: AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails

Create a value of AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deleteOnTermination:AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails', awsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails_deleteOnTermination - Whether to delete the volume when the instance is terminated.

$sel:volumeSize:AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails', awsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails_volumeSize - The volume size, in GiBs. The following are the supported volumes sizes for each volume type:

gp2 and gp3: 1-16,384

io1: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024

You must specify either SnapshotId or VolumeSize. If you specify both SnapshotId and VolumeSize, the volume size must be equal or greater than the size of the snapshot.

$sel:iops:AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails', awsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails_iops - The number of input/output (I/O) operations per second (IOPS) to provision for the volume.

Only supported for gp3 or io1 volumes. Required for io1 volumes. Not used with standard, gp2, st1, or sc1 volumes.

$sel:encrypted:AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails', awsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails_encrypted - Whether to encrypt the volume.

$sel:volumeType:AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails', awsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails_volumeType - The volume type.

$sel:snapshotId:AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails', awsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails_snapshotId - The snapshot ID of the volume to use.

You must specify either VolumeSize or SnapshotId.

awsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails_deleteOnTermination :: Lens' AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails (Maybe Bool)

Whether to delete the volume when the instance is terminated.

awsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails_volumeSize :: Lens' AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails (Maybe Int)

The volume size, in GiBs. The following are the supported volumes sizes for each volume type:

gp2 and gp3: 1-16,384

io1: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024

You must specify either SnapshotId or VolumeSize. If you specify both SnapshotId and VolumeSize, the volume size must be equal or greater than the size of the snapshot.

awsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails_iops :: Lens' AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails (Maybe Int)

The number of input/output (I/O) operations per second (IOPS) to provision for the volume.

Only supported for gp3 or io1 volumes. Required for io1 volumes. Not used with standard, gp2, st1, or sc1 volumes.

awsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails_encrypted :: Lens' AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails (Maybe Bool)

Whether to encrypt the volume.

awsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails_volumeType :: Lens' AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails (Maybe Text)

The volume type.

awsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails_snapshotId :: Lens' AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails (Maybe Text)

The snapshot ID of the volume to use.

You must specify either VolumeSize or SnapshotId.