Network.AWS.SecurityHub.Types.AwsAutoScalingLaunchConfigurationDetails

Description

Documentation

data AwsAutoScalingLaunchConfigurationDetails

Details about a launch configuration.

See: newAwsAutoScalingLaunchConfigurationDetails smart constructor.

Constructors

AwsAutoScalingLaunchConfigurationDetails'

Fields

associatePublicIpAddress :: Maybe Bool

For Auto Scaling groups that run in a VPC, specifies whether to assign a public IP address to the group's instances.

securityGroups :: Maybe [Text]

The security groups to assign to the instances in the Auto Scaling group.

spotPrice :: Maybe Text

The maximum hourly price to be paid for any Spot Instance that is launched to fulfill the request.

createdTime :: Maybe Text

The creation date and time for the launch configuration.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

instanceMonitoring :: Maybe AwsAutoScalingLaunchConfigurationInstanceMonitoringDetails

Indicates the type of monitoring for instances in the group.

keyName :: Maybe Text

The name of the key pair.

classicLinkVpcSecurityGroups :: Maybe [Text]

The identifiers of one or more security groups for the VPC that is specified in ClassicLinkVPCId.

ramdiskId :: Maybe Text

The identifier of the RAM disk associated with the AMI.

kernelId :: Maybe Text

The identifier of the kernel associated with the AMI.

instanceType :: Maybe Text

The instance type for the instances.

ebsOptimized :: Maybe Bool

Whether the launch configuration is optimized for Amazon EBS I/O.

userData :: Maybe Text

The user data to make available to the launched EC2 instances. Must be base64-encoded text.

classicLinkVpcId :: Maybe Text

The identifier of a ClassicLink-enabled VPC that EC2-Classic instances are linked to.

iamInstanceProfile :: Maybe Text

The name or the ARN of the instance profile associated with the IAM role for the instance. The instance profile contains the IAM role.

imageId :: Maybe Text

The identifier of the Amazon Machine Image (AMI) that is used to launch EC2 instances.

launchConfigurationName :: Maybe Text

The name of the launch configuration.

placementTenancy :: Maybe Text

The tenancy of the instance. An instance with dedicated tenancy runs on isolated, single-tenant hardware and can only be launched into a VPC.

blockDeviceMappings :: Maybe [AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails]

Specifies the block devices for the instance.

newAwsAutoScalingLaunchConfigurationDetails :: AwsAutoScalingLaunchConfigurationDetails

Create a value of AwsAutoScalingLaunchConfigurationDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associatePublicIpAddress:AwsAutoScalingLaunchConfigurationDetails', awsAutoScalingLaunchConfigurationDetails_associatePublicIpAddress - For Auto Scaling groups that run in a VPC, specifies whether to assign a public IP address to the group's instances.

$sel:securityGroups:AwsAutoScalingLaunchConfigurationDetails', awsAutoScalingLaunchConfigurationDetails_securityGroups - The security groups to assign to the instances in the Auto Scaling group.

$sel:spotPrice:AwsAutoScalingLaunchConfigurationDetails', awsAutoScalingLaunchConfigurationDetails_spotPrice - The maximum hourly price to be paid for any Spot Instance that is launched to fulfill the request.

$sel:createdTime:AwsAutoScalingLaunchConfigurationDetails', awsAutoScalingLaunchConfigurationDetails_createdTime - The creation date and time for the launch configuration.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:instanceMonitoring:AwsAutoScalingLaunchConfigurationDetails', awsAutoScalingLaunchConfigurationDetails_instanceMonitoring - Indicates the type of monitoring for instances in the group.

$sel:keyName:AwsAutoScalingLaunchConfigurationDetails', awsAutoScalingLaunchConfigurationDetails_keyName - The name of the key pair.

$sel:classicLinkVpcSecurityGroups:AwsAutoScalingLaunchConfigurationDetails', awsAutoScalingLaunchConfigurationDetails_classicLinkVpcSecurityGroups - The identifiers of one or more security groups for the VPC that is specified in ClassicLinkVPCId.

$sel:ramdiskId:AwsAutoScalingLaunchConfigurationDetails', awsAutoScalingLaunchConfigurationDetails_ramdiskId - The identifier of the RAM disk associated with the AMI.

$sel:kernelId:AwsAutoScalingLaunchConfigurationDetails', awsAutoScalingLaunchConfigurationDetails_kernelId - The identifier of the kernel associated with the AMI.

$sel:instanceType:AwsAutoScalingLaunchConfigurationDetails', awsAutoScalingLaunchConfigurationDetails_instanceType - The instance type for the instances.

$sel:ebsOptimized:AwsAutoScalingLaunchConfigurationDetails', awsAutoScalingLaunchConfigurationDetails_ebsOptimized - Whether the launch configuration is optimized for Amazon EBS I/O.

$sel:userData:AwsAutoScalingLaunchConfigurationDetails', awsAutoScalingLaunchConfigurationDetails_userData - The user data to make available to the launched EC2 instances. Must be base64-encoded text.

$sel:classicLinkVpcId:AwsAutoScalingLaunchConfigurationDetails', awsAutoScalingLaunchConfigurationDetails_classicLinkVpcId - The identifier of a ClassicLink-enabled VPC that EC2-Classic instances are linked to.

$sel:iamInstanceProfile:AwsAutoScalingLaunchConfigurationDetails', awsAutoScalingLaunchConfigurationDetails_iamInstanceProfile - The name or the ARN of the instance profile associated with the IAM role for the instance. The instance profile contains the IAM role.

$sel:imageId:AwsAutoScalingLaunchConfigurationDetails', awsAutoScalingLaunchConfigurationDetails_imageId - The identifier of the Amazon Machine Image (AMI) that is used to launch EC2 instances.

$sel:launchConfigurationName:AwsAutoScalingLaunchConfigurationDetails', awsAutoScalingLaunchConfigurationDetails_launchConfigurationName - The name of the launch configuration.

$sel:placementTenancy:AwsAutoScalingLaunchConfigurationDetails', awsAutoScalingLaunchConfigurationDetails_placementTenancy - The tenancy of the instance. An instance with dedicated tenancy runs on isolated, single-tenant hardware and can only be launched into a VPC.

$sel:blockDeviceMappings:AwsAutoScalingLaunchConfigurationDetails', awsAutoScalingLaunchConfigurationDetails_blockDeviceMappings - Specifies the block devices for the instance.

awsAutoScalingLaunchConfigurationDetails_associatePublicIpAddress :: Lens' AwsAutoScalingLaunchConfigurationDetails (Maybe Bool)

For Auto Scaling groups that run in a VPC, specifies whether to assign a public IP address to the group's instances.

awsAutoScalingLaunchConfigurationDetails_securityGroups :: Lens' AwsAutoScalingLaunchConfigurationDetails (Maybe [Text])

The security groups to assign to the instances in the Auto Scaling group.

awsAutoScalingLaunchConfigurationDetails_spotPrice :: Lens' AwsAutoScalingLaunchConfigurationDetails (Maybe Text)

The maximum hourly price to be paid for any Spot Instance that is launched to fulfill the request.

awsAutoScalingLaunchConfigurationDetails_createdTime :: Lens' AwsAutoScalingLaunchConfigurationDetails (Maybe Text)

The creation date and time for the launch configuration.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsAutoScalingLaunchConfigurationDetails_instanceMonitoring :: Lens' AwsAutoScalingLaunchConfigurationDetails (Maybe AwsAutoScalingLaunchConfigurationInstanceMonitoringDetails)

Indicates the type of monitoring for instances in the group.

awsAutoScalingLaunchConfigurationDetails_keyName :: Lens' AwsAutoScalingLaunchConfigurationDetails (Maybe Text)

The name of the key pair.

awsAutoScalingLaunchConfigurationDetails_classicLinkVpcSecurityGroups :: Lens' AwsAutoScalingLaunchConfigurationDetails (Maybe [Text])

The identifiers of one or more security groups for the VPC that is specified in ClassicLinkVPCId.

awsAutoScalingLaunchConfigurationDetails_ramdiskId :: Lens' AwsAutoScalingLaunchConfigurationDetails (Maybe Text)

The identifier of the RAM disk associated with the AMI.

awsAutoScalingLaunchConfigurationDetails_kernelId :: Lens' AwsAutoScalingLaunchConfigurationDetails (Maybe Text)

The identifier of the kernel associated with the AMI.

awsAutoScalingLaunchConfigurationDetails_instanceType :: Lens' AwsAutoScalingLaunchConfigurationDetails (Maybe Text)

The instance type for the instances.

awsAutoScalingLaunchConfigurationDetails_ebsOptimized :: Lens' AwsAutoScalingLaunchConfigurationDetails (Maybe Bool)

Whether the launch configuration is optimized for Amazon EBS I/O.

awsAutoScalingLaunchConfigurationDetails_userData :: Lens' AwsAutoScalingLaunchConfigurationDetails (Maybe Text)

The user data to make available to the launched EC2 instances. Must be base64-encoded text.

awsAutoScalingLaunchConfigurationDetails_classicLinkVpcId :: Lens' AwsAutoScalingLaunchConfigurationDetails (Maybe Text)

The identifier of a ClassicLink-enabled VPC that EC2-Classic instances are linked to.

awsAutoScalingLaunchConfigurationDetails_iamInstanceProfile :: Lens' AwsAutoScalingLaunchConfigurationDetails (Maybe Text)

The name or the ARN of the instance profile associated with the IAM role for the instance. The instance profile contains the IAM role.

awsAutoScalingLaunchConfigurationDetails_imageId :: Lens' AwsAutoScalingLaunchConfigurationDetails (Maybe Text)

The identifier of the Amazon Machine Image (AMI) that is used to launch EC2 instances.

awsAutoScalingLaunchConfigurationDetails_launchConfigurationName :: Lens' AwsAutoScalingLaunchConfigurationDetails (Maybe Text)

The name of the launch configuration.

awsAutoScalingLaunchConfigurationDetails_placementTenancy :: Lens' AwsAutoScalingLaunchConfigurationDetails (Maybe Text)

The tenancy of the instance. An instance with dedicated tenancy runs on isolated, single-tenant hardware and can only be launched into a VPC.

awsAutoScalingLaunchConfigurationDetails_blockDeviceMappings :: Lens' AwsAutoScalingLaunchConfigurationDetails (Maybe [AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails])

Specifies the block devices for the instance.