Network.AWS.SecurityHub.Types.AwsCertificateManagerCertificateDomainValidationOption

Description

Documentation

data AwsCertificateManagerCertificateDomainValidationOption

Contains information about one of the following:

The initial validation of each domain name that occurs as a result of the RequestCertificate request

The validation of each domain name in the certificate, as it pertains to Certificate Manager managed renewal

See: newAwsCertificateManagerCertificateDomainValidationOption smart constructor.

Constructors

AwsCertificateManagerCertificateDomainValidationOption'

Fields

validationEmails :: Maybe [Text]

A list of email addresses that Certificate Manager uses to send domain validation emails.

validationMethod :: Maybe Text

The method used to validate the domain name.

resourceRecord :: Maybe AwsCertificateManagerCertificateResourceRecord

The CNAME record that is added to the DNS database for domain validation.

validationStatus :: Maybe Text

The validation status of the domain name.

domainName :: Maybe Text

A fully qualified domain name (FQDN) in the certificate.

validationDomain :: Maybe Text

The domain name that Certificate Manager uses to send domain validation emails.

newAwsCertificateManagerCertificateDomainValidationOption :: AwsCertificateManagerCertificateDomainValidationOption

Create a value of AwsCertificateManagerCertificateDomainValidationOption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:validationEmails:AwsCertificateManagerCertificateDomainValidationOption', awsCertificateManagerCertificateDomainValidationOption_validationEmails - A list of email addresses that Certificate Manager uses to send domain validation emails.

$sel:validationMethod:AwsCertificateManagerCertificateDomainValidationOption', awsCertificateManagerCertificateDomainValidationOption_validationMethod - The method used to validate the domain name.

$sel:resourceRecord:AwsCertificateManagerCertificateDomainValidationOption', awsCertificateManagerCertificateDomainValidationOption_resourceRecord - The CNAME record that is added to the DNS database for domain validation.

$sel:validationStatus:AwsCertificateManagerCertificateDomainValidationOption', awsCertificateManagerCertificateDomainValidationOption_validationStatus - The validation status of the domain name.

$sel:domainName:AwsCertificateManagerCertificateDomainValidationOption', awsCertificateManagerCertificateDomainValidationOption_domainName - A fully qualified domain name (FQDN) in the certificate.

$sel:validationDomain:AwsCertificateManagerCertificateDomainValidationOption', awsCertificateManagerCertificateDomainValidationOption_validationDomain - The domain name that Certificate Manager uses to send domain validation emails.

awsCertificateManagerCertificateDomainValidationOption_validationEmails :: Lens' AwsCertificateManagerCertificateDomainValidationOption (Maybe [Text])

A list of email addresses that Certificate Manager uses to send domain validation emails.

awsCertificateManagerCertificateDomainValidationOption_validationMethod :: Lens' AwsCertificateManagerCertificateDomainValidationOption (Maybe Text)

The method used to validate the domain name.

awsCertificateManagerCertificateDomainValidationOption_resourceRecord :: Lens' AwsCertificateManagerCertificateDomainValidationOption (Maybe AwsCertificateManagerCertificateResourceRecord)

The CNAME record that is added to the DNS database for domain validation.

awsCertificateManagerCertificateDomainValidationOption_validationStatus :: Lens' AwsCertificateManagerCertificateDomainValidationOption (Maybe Text)

The validation status of the domain name.

awsCertificateManagerCertificateDomainValidationOption_domainName :: Lens' AwsCertificateManagerCertificateDomainValidationOption (Maybe Text)

A fully qualified domain name (FQDN) in the certificate.

awsCertificateManagerCertificateDomainValidationOption_validationDomain :: Lens' AwsCertificateManagerCertificateDomainValidationOption (Maybe Text)

The domain name that Certificate Manager uses to send domain validation emails.