Network.AWS.SecurityHub.Types.AwsCloudFrontDistributionDetails

Description

Documentation

data AwsCloudFrontDistributionDetails

A distribution configuration.

See: newAwsCloudFrontDistributionDetails smart constructor.

Constructors

AwsCloudFrontDistributionDetails'

Fields

status :: Maybe Text

Indicates the current status of the distribution.

eTag :: Maybe Text

The entity tag is a hash of the object.

originGroups :: Maybe AwsCloudFrontDistributionOriginGroups

Provides information about the origin groups in the distribution.

defaultRootObject :: Maybe Text

The object that CloudFront sends in response to requests from the origin (for example, index.html) when a viewer requests the root URL for the distribution (http://www.example.com) instead of an object in your distribution (http://www.example.com/product-description.html).

webAclId :: Maybe Text

A unique identifier that specifies the WAF web ACL, if any, to associate with this distribution.

lastModifiedTime :: Maybe Text

Indicates when that the distribution was last modified.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

viewerCertificate :: Maybe AwsCloudFrontDistributionViewerCertificate

Provides information about the TLS/SSL configuration that the distribution uses to communicate with viewers.

domainName :: Maybe Text

The domain name corresponding to the distribution.

origins :: Maybe AwsCloudFrontDistributionOrigins

A complex type that contains information about origins for this distribution.

logging :: Maybe AwsCloudFrontDistributionLogging

A complex type that controls whether access logs are written for the distribution.

cacheBehaviors :: Maybe AwsCloudFrontDistributionCacheBehaviors

Provides information about the cache configuration for the distribution.

defaultCacheBehavior :: Maybe AwsCloudFrontDistributionDefaultCacheBehavior

The default cache behavior for the configuration.

newAwsCloudFrontDistributionDetails :: AwsCloudFrontDistributionDetails

Create a value of AwsCloudFrontDistributionDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AwsCloudFrontDistributionDetails', awsCloudFrontDistributionDetails_status - Indicates the current status of the distribution.

$sel:eTag:AwsCloudFrontDistributionDetails', awsCloudFrontDistributionDetails_eTag - The entity tag is a hash of the object.

$sel:originGroups:AwsCloudFrontDistributionDetails', awsCloudFrontDistributionDetails_originGroups - Provides information about the origin groups in the distribution.

$sel:defaultRootObject:AwsCloudFrontDistributionDetails', awsCloudFrontDistributionDetails_defaultRootObject - The object that CloudFront sends in response to requests from the origin (for example, index.html) when a viewer requests the root URL for the distribution (http://www.example.com) instead of an object in your distribution (http://www.example.com/product-description.html).

$sel:webAclId:AwsCloudFrontDistributionDetails', awsCloudFrontDistributionDetails_webAclId - A unique identifier that specifies the WAF web ACL, if any, to associate with this distribution.

$sel:lastModifiedTime:AwsCloudFrontDistributionDetails', awsCloudFrontDistributionDetails_lastModifiedTime - Indicates when that the distribution was last modified.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:viewerCertificate:AwsCloudFrontDistributionDetails', awsCloudFrontDistributionDetails_viewerCertificate - Provides information about the TLS/SSL configuration that the distribution uses to communicate with viewers.

$sel:domainName:AwsCloudFrontDistributionDetails', awsCloudFrontDistributionDetails_domainName - The domain name corresponding to the distribution.

$sel:origins:AwsCloudFrontDistributionDetails', awsCloudFrontDistributionDetails_origins - A complex type that contains information about origins for this distribution.

$sel:logging:AwsCloudFrontDistributionDetails', awsCloudFrontDistributionDetails_logging - A complex type that controls whether access logs are written for the distribution.

$sel:cacheBehaviors:AwsCloudFrontDistributionDetails', awsCloudFrontDistributionDetails_cacheBehaviors - Provides information about the cache configuration for the distribution.

$sel:defaultCacheBehavior:AwsCloudFrontDistributionDetails', awsCloudFrontDistributionDetails_defaultCacheBehavior - The default cache behavior for the configuration.

awsCloudFrontDistributionDetails_status :: Lens' AwsCloudFrontDistributionDetails (Maybe Text)

Indicates the current status of the distribution.

awsCloudFrontDistributionDetails_eTag :: Lens' AwsCloudFrontDistributionDetails (Maybe Text)

The entity tag is a hash of the object.

awsCloudFrontDistributionDetails_originGroups :: Lens' AwsCloudFrontDistributionDetails (Maybe AwsCloudFrontDistributionOriginGroups)

Provides information about the origin groups in the distribution.

awsCloudFrontDistributionDetails_defaultRootObject :: Lens' AwsCloudFrontDistributionDetails (Maybe Text)

The object that CloudFront sends in response to requests from the origin (for example, index.html) when a viewer requests the root URL for the distribution (http://www.example.com) instead of an object in your distribution (http://www.example.com/product-description.html).

awsCloudFrontDistributionDetails_webAclId :: Lens' AwsCloudFrontDistributionDetails (Maybe Text)

A unique identifier that specifies the WAF web ACL, if any, to associate with this distribution.

awsCloudFrontDistributionDetails_lastModifiedTime :: Lens' AwsCloudFrontDistributionDetails (Maybe Text)

Indicates when that the distribution was last modified.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsCloudFrontDistributionDetails_viewerCertificate :: Lens' AwsCloudFrontDistributionDetails (Maybe AwsCloudFrontDistributionViewerCertificate)

Provides information about the TLS/SSL configuration that the distribution uses to communicate with viewers.

awsCloudFrontDistributionDetails_domainName :: Lens' AwsCloudFrontDistributionDetails (Maybe Text)

The domain name corresponding to the distribution.

awsCloudFrontDistributionDetails_origins :: Lens' AwsCloudFrontDistributionDetails (Maybe AwsCloudFrontDistributionOrigins)

A complex type that contains information about origins for this distribution.

awsCloudFrontDistributionDetails_logging :: Lens' AwsCloudFrontDistributionDetails (Maybe AwsCloudFrontDistributionLogging)

A complex type that controls whether access logs are written for the distribution.

awsCloudFrontDistributionDetails_cacheBehaviors :: Lens' AwsCloudFrontDistributionDetails (Maybe AwsCloudFrontDistributionCacheBehaviors)

Provides information about the cache configuration for the distribution.

awsCloudFrontDistributionDetails_defaultCacheBehavior :: Lens' AwsCloudFrontDistributionDetails (Maybe AwsCloudFrontDistributionDefaultCacheBehavior)

The default cache behavior for the configuration.