Network.AWS.SecurityHub.Types.AwsCodeBuildProjectEnvironment

Description

Documentation

data AwsCodeBuildProjectEnvironment

Information about the build environment for this build project.

See: newAwsCodeBuildProjectEnvironment smart constructor.

Constructors

AwsCodeBuildProjectEnvironment'

Fields

imagePullCredentialsType :: Maybe Text

The type of credentials CodeBuild uses to pull images in your build.

Valid values:

CODEBUILD specifies that CodeBuild uses its own credentials. This requires that you modify your ECR repository policy to trust the CodeBuild service principal.

SERVICE_ROLE specifies that CodeBuild uses your build project's service role.

When you use a cross-account or private registry image, you must use SERVICE_ROLE credentials. When you use an CodeBuild curated image, you must use CODEBUILD credentials.

privilegedMode :: Maybe Bool

Whether to allow the Docker daemon to run inside a Docker container. Set to true if the build project is used to build Docker images.

registryCredential :: Maybe AwsCodeBuildProjectEnvironmentRegistryCredential

The credentials for access to a private registry.

certificate :: Maybe Text

The certificate to use with this build project.

environmentVariables :: Maybe [AwsCodeBuildProjectEnvironmentEnvironmentVariablesDetails]

A set of environment variables to make available to builds for the build project.

type' :: Maybe Text

The type of build environment to use for related builds.

The environment type ARM_CONTAINER is available only in Regions US East (N. Virginia), US East (Ohio), US West (Oregon), Europe (Ireland), Asia Pacific (Mumbai), Asia Pacific (Tokyo), Asia Pacific (Sydney), and Europe (Frankfurt).

The environment type LINUX_CONTAINER with compute type build.general1.2xlarge is available only in Regions US East (N. Virginia), US East (N. Virginia), US West (Oregon), Canada (Central), Europe (Ireland), Europe (London), Europe (Frankfurt), Asia Pacific (Tokyo), Asia Pacific (Seoul), Asia Pacific (Singapore), Asia Pacific (Sydney), China (Beijing), and China (Ningxia).

The environment type LINUX_GPU_CONTAINER is available only in Regions US East (N. Virginia), US East (N. Virginia), US West (Oregon), Canada (Central), Europe (Ireland), Europe (London), Europe (Frankfurt), Asia Pacific (Tokyo), Asia Pacific (Seoul), Asia Pacific (Singapore), Asia Pacific (Sydney), China (Beijing), and China (Ningxia).

Valid values: WINDOWS_CONTAINER | LINUX_CONTAINER | LINUX_GPU_CONTAINER | ARM_CONTAINER

newAwsCodeBuildProjectEnvironment :: AwsCodeBuildProjectEnvironment

Create a value of AwsCodeBuildProjectEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imagePullCredentialsType:AwsCodeBuildProjectEnvironment', awsCodeBuildProjectEnvironment_imagePullCredentialsType - The type of credentials CodeBuild uses to pull images in your build.

Valid values:

CODEBUILD specifies that CodeBuild uses its own credentials. This requires that you modify your ECR repository policy to trust the CodeBuild service principal.

SERVICE_ROLE specifies that CodeBuild uses your build project's service role.

When you use a cross-account or private registry image, you must use SERVICE_ROLE credentials. When you use an CodeBuild curated image, you must use CODEBUILD credentials.

$sel:privilegedMode:AwsCodeBuildProjectEnvironment', awsCodeBuildProjectEnvironment_privilegedMode - Whether to allow the Docker daemon to run inside a Docker container. Set to true if the build project is used to build Docker images.

$sel:registryCredential:AwsCodeBuildProjectEnvironment', awsCodeBuildProjectEnvironment_registryCredential - The credentials for access to a private registry.

$sel:certificate:AwsCodeBuildProjectEnvironment', awsCodeBuildProjectEnvironment_certificate - The certificate to use with this build project.

$sel:environmentVariables:AwsCodeBuildProjectEnvironment', awsCodeBuildProjectEnvironment_environmentVariables - A set of environment variables to make available to builds for the build project.

$sel:type':AwsCodeBuildProjectEnvironment', awsCodeBuildProjectEnvironment_type - The type of build environment to use for related builds.

The environment type ARM_CONTAINER is available only in Regions US East (N. Virginia), US East (Ohio), US West (Oregon), Europe (Ireland), Asia Pacific (Mumbai), Asia Pacific (Tokyo), Asia Pacific (Sydney), and Europe (Frankfurt).

The environment type LINUX_CONTAINER with compute type build.general1.2xlarge is available only in Regions US East (N. Virginia), US East (N. Virginia), US West (Oregon), Canada (Central), Europe (Ireland), Europe (London), Europe (Frankfurt), Asia Pacific (Tokyo), Asia Pacific (Seoul), Asia Pacific (Singapore), Asia Pacific (Sydney), China (Beijing), and China (Ningxia).

The environment type LINUX_GPU_CONTAINER is available only in Regions US East (N. Virginia), US East (N. Virginia), US West (Oregon), Canada (Central), Europe (Ireland), Europe (London), Europe (Frankfurt), Asia Pacific (Tokyo), Asia Pacific (Seoul), Asia Pacific (Singapore), Asia Pacific (Sydney), China (Beijing), and China (Ningxia).

Valid values: WINDOWS_CONTAINER | LINUX_CONTAINER | LINUX_GPU_CONTAINER | ARM_CONTAINER

awsCodeBuildProjectEnvironment_imagePullCredentialsType :: Lens' AwsCodeBuildProjectEnvironment (Maybe Text)

The type of credentials CodeBuild uses to pull images in your build.

Valid values:

CODEBUILD specifies that CodeBuild uses its own credentials. This requires that you modify your ECR repository policy to trust the CodeBuild service principal.

SERVICE_ROLE specifies that CodeBuild uses your build project's service role.

When you use a cross-account or private registry image, you must use SERVICE_ROLE credentials. When you use an CodeBuild curated image, you must use CODEBUILD credentials.

awsCodeBuildProjectEnvironment_privilegedMode :: Lens' AwsCodeBuildProjectEnvironment (Maybe Bool)

Whether to allow the Docker daemon to run inside a Docker container. Set to true if the build project is used to build Docker images.

awsCodeBuildProjectEnvironment_registryCredential :: Lens' AwsCodeBuildProjectEnvironment (Maybe AwsCodeBuildProjectEnvironmentRegistryCredential)

The credentials for access to a private registry.

awsCodeBuildProjectEnvironment_certificate :: Lens' AwsCodeBuildProjectEnvironment (Maybe Text)

The certificate to use with this build project.

awsCodeBuildProjectEnvironment_environmentVariables :: Lens' AwsCodeBuildProjectEnvironment (Maybe [AwsCodeBuildProjectEnvironmentEnvironmentVariablesDetails])

A set of environment variables to make available to builds for the build project.

awsCodeBuildProjectEnvironment_type :: Lens' AwsCodeBuildProjectEnvironment (Maybe Text)

The type of build environment to use for related builds.

The environment type ARM_CONTAINER is available only in Regions US East (N. Virginia), US East (Ohio), US West (Oregon), Europe (Ireland), Asia Pacific (Mumbai), Asia Pacific (Tokyo), Asia Pacific (Sydney), and Europe (Frankfurt).

The environment type LINUX_CONTAINER with compute type build.general1.2xlarge is available only in Regions US East (N. Virginia), US East (N. Virginia), US West (Oregon), Canada (Central), Europe (Ireland), Europe (London), Europe (Frankfurt), Asia Pacific (Tokyo), Asia Pacific (Seoul), Asia Pacific (Singapore), Asia Pacific (Sydney), China (Beijing), and China (Ningxia).

The environment type LINUX_GPU_CONTAINER is available only in Regions US East (N. Virginia), US East (N. Virginia), US West (Oregon), Canada (Central), Europe (Ireland), Europe (London), Europe (Frankfurt), Asia Pacific (Tokyo), Asia Pacific (Seoul), Asia Pacific (Singapore), Asia Pacific (Sydney), China (Beijing), and China (Ningxia).

Valid values: WINDOWS_CONTAINER | LINUX_CONTAINER | LINUX_GPU_CONTAINER | ARM_CONTAINER