Network.AWS.SecurityHub.Types.AwsDynamoDbTableProjection

Description

Documentation

data AwsDynamoDbTableProjection

For global and local secondary indexes, identifies the attributes that are copied from the table into the index.

See: newAwsDynamoDbTableProjection smart constructor.

Constructors

AwsDynamoDbTableProjection'

Fields

projectionType :: Maybe Text

The types of attributes that are projected into the index.

nonKeyAttributes :: Maybe [Text]

The nonkey attributes that are projected into the index. For each attribute, provide the attribute name.

newAwsDynamoDbTableProjection :: AwsDynamoDbTableProjection

Create a value of AwsDynamoDbTableProjection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectionType:AwsDynamoDbTableProjection', awsDynamoDbTableProjection_projectionType - The types of attributes that are projected into the index.

$sel:nonKeyAttributes:AwsDynamoDbTableProjection', awsDynamoDbTableProjection_nonKeyAttributes - The nonkey attributes that are projected into the index. For each attribute, provide the attribute name.

awsDynamoDbTableProjection_projectionType :: Lens' AwsDynamoDbTableProjection (Maybe Text)

The types of attributes that are projected into the index.

awsDynamoDbTableProjection_nonKeyAttributes :: Lens' AwsDynamoDbTableProjection (Maybe [Text])

The nonkey attributes that are projected into the index. For each attribute, provide the attribute name.