Network.AWS.SecurityHub.Types.AwsDynamoDbTableReplica

Description

Documentation

data AwsDynamoDbTableReplica

Information about a replica of a DynamoDB table.

See: newAwsDynamoDbTableReplica smart constructor.

Constructors

AwsDynamoDbTableReplica'

Fields

replicaStatus :: Maybe Text

The current status of the replica.

regionName :: Maybe Text

The name of the Region where the replica is located.

replicaStatusDescription :: Maybe Text

Detailed information about the replica status.

kmsMasterKeyId :: Maybe Text

The identifier of the KMS key that will be used for KMS encryption for the replica.

provisionedThroughputOverride :: Maybe AwsDynamoDbTableProvisionedThroughputOverride

Replica-specific configuration for the provisioned throughput.

globalSecondaryIndexes :: Maybe [AwsDynamoDbTableReplicaGlobalSecondaryIndex]

List of global secondary indexes for the replica.

newAwsDynamoDbTableReplica :: AwsDynamoDbTableReplica

Create a value of AwsDynamoDbTableReplica with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replicaStatus:AwsDynamoDbTableReplica', awsDynamoDbTableReplica_replicaStatus - The current status of the replica.

$sel:regionName:AwsDynamoDbTableReplica', awsDynamoDbTableReplica_regionName - The name of the Region where the replica is located.

$sel:replicaStatusDescription:AwsDynamoDbTableReplica', awsDynamoDbTableReplica_replicaStatusDescription - Detailed information about the replica status.

$sel:kmsMasterKeyId:AwsDynamoDbTableReplica', awsDynamoDbTableReplica_kmsMasterKeyId - The identifier of the KMS key that will be used for KMS encryption for the replica.

$sel:provisionedThroughputOverride:AwsDynamoDbTableReplica', awsDynamoDbTableReplica_provisionedThroughputOverride - Replica-specific configuration for the provisioned throughput.

$sel:globalSecondaryIndexes:AwsDynamoDbTableReplica', awsDynamoDbTableReplica_globalSecondaryIndexes - List of global secondary indexes for the replica.

awsDynamoDbTableReplica_replicaStatus :: Lens' AwsDynamoDbTableReplica (Maybe Text)

The current status of the replica.

awsDynamoDbTableReplica_regionName :: Lens' AwsDynamoDbTableReplica (Maybe Text)

The name of the Region where the replica is located.

awsDynamoDbTableReplica_replicaStatusDescription :: Lens' AwsDynamoDbTableReplica (Maybe Text)

Detailed information about the replica status.

awsDynamoDbTableReplica_kmsMasterKeyId :: Lens' AwsDynamoDbTableReplica (Maybe Text)

The identifier of the KMS key that will be used for KMS encryption for the replica.

awsDynamoDbTableReplica_provisionedThroughputOverride :: Lens' AwsDynamoDbTableReplica (Maybe AwsDynamoDbTableProvisionedThroughputOverride)

Replica-specific configuration for the provisioned throughput.

awsDynamoDbTableReplica_globalSecondaryIndexes :: Lens' AwsDynamoDbTableReplica (Maybe [AwsDynamoDbTableReplicaGlobalSecondaryIndex])

List of global secondary indexes for the replica.