Network.AWS.SecurityHub.Types.AwsDynamoDbTableRestoreSummary

Description

Documentation

data AwsDynamoDbTableRestoreSummary

Information about the restore for the table.

See: newAwsDynamoDbTableRestoreSummary smart constructor.

Constructors

AwsDynamoDbTableRestoreSummary'

Fields

sourceTableArn :: Maybe Text

The ARN of the source table for the backup.

sourceBackupArn :: Maybe Text

The ARN of the source backup from which the table was restored.

restoreInProgress :: Maybe Bool

Whether a restore is currently in progress.

restoreDateTime :: Maybe Text

Indicates the point in time that the table was restored to.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

newAwsDynamoDbTableRestoreSummary :: AwsDynamoDbTableRestoreSummary

Create a value of AwsDynamoDbTableRestoreSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceTableArn:AwsDynamoDbTableRestoreSummary', awsDynamoDbTableRestoreSummary_sourceTableArn - The ARN of the source table for the backup.

$sel:sourceBackupArn:AwsDynamoDbTableRestoreSummary', awsDynamoDbTableRestoreSummary_sourceBackupArn - The ARN of the source backup from which the table was restored.

$sel:restoreInProgress:AwsDynamoDbTableRestoreSummary', awsDynamoDbTableRestoreSummary_restoreInProgress - Whether a restore is currently in progress.

$sel:restoreDateTime:AwsDynamoDbTableRestoreSummary', awsDynamoDbTableRestoreSummary_restoreDateTime - Indicates the point in time that the table was restored to.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsDynamoDbTableRestoreSummary_sourceTableArn :: Lens' AwsDynamoDbTableRestoreSummary (Maybe Text)

The ARN of the source table for the backup.

awsDynamoDbTableRestoreSummary_sourceBackupArn :: Lens' AwsDynamoDbTableRestoreSummary (Maybe Text)

The ARN of the source backup from which the table was restored.

awsDynamoDbTableRestoreSummary_restoreInProgress :: Lens' AwsDynamoDbTableRestoreSummary (Maybe Bool)

Whether a restore is currently in progress.

awsDynamoDbTableRestoreSummary_restoreDateTime :: Lens' AwsDynamoDbTableRestoreSummary (Maybe Text)

Indicates the point in time that the table was restored to.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.