Network.AWS.SecurityHub.Types.AwsEc2EipDetails

Description

Documentation

data AwsEc2EipDetails

Information about an Elastic IP address.

See: newAwsEc2EipDetails smart constructor.

Constructors

AwsEc2EipDetails'

Fields

associationId :: Maybe Text

The identifier that represents the association of the Elastic IP address with an EC2 instance.

instanceId :: Maybe Text

The identifier of the EC2 instance.

networkInterfaceOwnerId :: Maybe Text

The Amazon Web Services account ID of the owner of the network interface.

allocationId :: Maybe Text

The identifier that Amazon Web Services assigns to represent the allocation of the Elastic IP address for use with Amazon VPC.

networkBorderGroup :: Maybe Text

The name of the location from which the Elastic IP address is advertised.

domain :: Maybe Text

The domain in which to allocate the address.

If the address is for use with EC2 instances in a VPC, then Domain is vpc. Otherwise, Domain is standard.

networkInterfaceId :: Maybe Text

The identifier of the network interface.

publicIpv4Pool :: Maybe Text

The identifier of an IP address pool. This parameter allows Amazon EC2 to select an IP address from the address pool.

privateIpAddress :: Maybe Text

The private IP address that is associated with the Elastic IP address.

publicIp :: Maybe Text

A public IP address that is associated with the EC2 instance.

newAwsEc2EipDetails :: AwsEc2EipDetails

Create a value of AwsEc2EipDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associationId:AwsEc2EipDetails', awsEc2EipDetails_associationId - The identifier that represents the association of the Elastic IP address with an EC2 instance.

$sel:instanceId:AwsEc2EipDetails', awsEc2EipDetails_instanceId - The identifier of the EC2 instance.

$sel:networkInterfaceOwnerId:AwsEc2EipDetails', awsEc2EipDetails_networkInterfaceOwnerId - The Amazon Web Services account ID of the owner of the network interface.

$sel:allocationId:AwsEc2EipDetails', awsEc2EipDetails_allocationId - The identifier that Amazon Web Services assigns to represent the allocation of the Elastic IP address for use with Amazon VPC.

$sel:networkBorderGroup:AwsEc2EipDetails', awsEc2EipDetails_networkBorderGroup - The name of the location from which the Elastic IP address is advertised.

$sel:domain:AwsEc2EipDetails', awsEc2EipDetails_domain - The domain in which to allocate the address.

If the address is for use with EC2 instances in a VPC, then Domain is vpc. Otherwise, Domain is standard.

$sel:networkInterfaceId:AwsEc2EipDetails', awsEc2EipDetails_networkInterfaceId - The identifier of the network interface.

$sel:publicIpv4Pool:AwsEc2EipDetails', awsEc2EipDetails_publicIpv4Pool - The identifier of an IP address pool. This parameter allows Amazon EC2 to select an IP address from the address pool.

$sel:privateIpAddress:AwsEc2EipDetails', awsEc2EipDetails_privateIpAddress - The private IP address that is associated with the Elastic IP address.

$sel:publicIp:AwsEc2EipDetails', awsEc2EipDetails_publicIp - A public IP address that is associated with the EC2 instance.

awsEc2EipDetails_associationId :: Lens' AwsEc2EipDetails (Maybe Text)

The identifier that represents the association of the Elastic IP address with an EC2 instance.

awsEc2EipDetails_instanceId :: Lens' AwsEc2EipDetails (Maybe Text)

The identifier of the EC2 instance.

awsEc2EipDetails_networkInterfaceOwnerId :: Lens' AwsEc2EipDetails (Maybe Text)

The Amazon Web Services account ID of the owner of the network interface.

awsEc2EipDetails_allocationId :: Lens' AwsEc2EipDetails (Maybe Text)

The identifier that Amazon Web Services assigns to represent the allocation of the Elastic IP address for use with Amazon VPC.

awsEc2EipDetails_networkBorderGroup :: Lens' AwsEc2EipDetails (Maybe Text)

The name of the location from which the Elastic IP address is advertised.

awsEc2EipDetails_domain :: Lens' AwsEc2EipDetails (Maybe Text)

The domain in which to allocate the address.

If the address is for use with EC2 instances in a VPC, then Domain is vpc. Otherwise, Domain is standard.

awsEc2EipDetails_networkInterfaceId :: Lens' AwsEc2EipDetails (Maybe Text)

The identifier of the network interface.

awsEc2EipDetails_publicIpv4Pool :: Lens' AwsEc2EipDetails (Maybe Text)

The identifier of an IP address pool. This parameter allows Amazon EC2 to select an IP address from the address pool.

awsEc2EipDetails_privateIpAddress :: Lens' AwsEc2EipDetails (Maybe Text)

The private IP address that is associated with the Elastic IP address.

awsEc2EipDetails_publicIp :: Lens' AwsEc2EipDetails (Maybe Text)

A public IP address that is associated with the EC2 instance.