Network.AWS.SecurityHub.Types.AwsEc2NetworkAclEntry

Description

Documentation

data AwsEc2NetworkAclEntry

A rule for the network ACL. Each rule allows or denies access based on the IP address, traffic direction, port, and protocol.

See: newAwsEc2NetworkAclEntry smart constructor.

Constructors

AwsEc2NetworkAclEntry'

Fields

ipv6CidrBlock :: Maybe Text

The IPV6 network range for which to deny or allow access.

icmpTypeCode :: Maybe IcmpTypeCode

The Internet Control Message Protocol (ICMP) type and code for which to deny or allow access.

ruleNumber :: Maybe Int

The rule number. The rules are processed in order by their number.

ruleAction :: Maybe Text

Whether the rule is used to allow access or deny access.

protocol :: Maybe Text

The protocol that the rule applies to. To deny or allow access to all protocols, use the value -1.

portRange :: Maybe PortRangeFromTo

For TCP or UDP protocols, the range of ports that the rule applies to.

cidrBlock :: Maybe Text

The IPV4 network range for which to deny or allow access.

egress :: Maybe Bool

Whether the rule is an egress rule. An egress rule is a rule that applies to traffic that leaves the subnet.

newAwsEc2NetworkAclEntry :: AwsEc2NetworkAclEntry

Create a value of AwsEc2NetworkAclEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipv6CidrBlock:AwsEc2NetworkAclEntry', awsEc2NetworkAclEntry_ipv6CidrBlock - The IPV6 network range for which to deny or allow access.

$sel:icmpTypeCode:AwsEc2NetworkAclEntry', awsEc2NetworkAclEntry_icmpTypeCode - The Internet Control Message Protocol (ICMP) type and code for which to deny or allow access.

$sel:ruleNumber:AwsEc2NetworkAclEntry', awsEc2NetworkAclEntry_ruleNumber - The rule number. The rules are processed in order by their number.

$sel:ruleAction:AwsEc2NetworkAclEntry', awsEc2NetworkAclEntry_ruleAction - Whether the rule is used to allow access or deny access.

$sel:protocol:AwsEc2NetworkAclEntry', awsEc2NetworkAclEntry_protocol - The protocol that the rule applies to. To deny or allow access to all protocols, use the value -1.

$sel:portRange:AwsEc2NetworkAclEntry', awsEc2NetworkAclEntry_portRange - For TCP or UDP protocols, the range of ports that the rule applies to.

$sel:cidrBlock:AwsEc2NetworkAclEntry', awsEc2NetworkAclEntry_cidrBlock - The IPV4 network range for which to deny or allow access.

$sel:egress:AwsEc2NetworkAclEntry', awsEc2NetworkAclEntry_egress - Whether the rule is an egress rule. An egress rule is a rule that applies to traffic that leaves the subnet.

awsEc2NetworkAclEntry_ipv6CidrBlock :: Lens' AwsEc2NetworkAclEntry (Maybe Text)

The IPV6 network range for which to deny or allow access.

awsEc2NetworkAclEntry_icmpTypeCode :: Lens' AwsEc2NetworkAclEntry (Maybe IcmpTypeCode)

The Internet Control Message Protocol (ICMP) type and code for which to deny or allow access.

awsEc2NetworkAclEntry_ruleNumber :: Lens' AwsEc2NetworkAclEntry (Maybe Int)

The rule number. The rules are processed in order by their number.

awsEc2NetworkAclEntry_ruleAction :: Lens' AwsEc2NetworkAclEntry (Maybe Text)

Whether the rule is used to allow access or deny access.

awsEc2NetworkAclEntry_protocol :: Lens' AwsEc2NetworkAclEntry (Maybe Text)

The protocol that the rule applies to. To deny or allow access to all protocols, use the value -1.

awsEc2NetworkAclEntry_portRange :: Lens' AwsEc2NetworkAclEntry (Maybe PortRangeFromTo)

For TCP or UDP protocols, the range of ports that the rule applies to.

awsEc2NetworkAclEntry_cidrBlock :: Lens' AwsEc2NetworkAclEntry (Maybe Text)

The IPV4 network range for which to deny or allow access.

awsEc2NetworkAclEntry_egress :: Lens' AwsEc2NetworkAclEntry (Maybe Bool)

Whether the rule is an egress rule. An egress rule is a rule that applies to traffic that leaves the subnet.