Network.AWS.SecurityHub.Types.AwsEc2NetworkInterfaceAttachment

Description

Documentation

data AwsEc2NetworkInterfaceAttachment

Information about the network interface attachment.

See: newAwsEc2NetworkInterfaceAttachment smart constructor.

Constructors

AwsEc2NetworkInterfaceAttachment'

Fields

instanceId :: Maybe Text

The ID of the instance.

status :: Maybe Text

The attachment state.

Valid values: attaching | attached | detaching | detached

deleteOnTermination :: Maybe Bool

Indicates whether the network interface is deleted when the instance is terminated.

attachmentId :: Maybe Text

The identifier of the network interface attachment

instanceOwnerId :: Maybe Text

The Amazon Web Services account ID of the owner of the instance.

attachTime :: Maybe Text

Indicates when the attachment initiated.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

deviceIndex :: Maybe Int

The device index of the network interface attachment on the instance.

newAwsEc2NetworkInterfaceAttachment :: AwsEc2NetworkInterfaceAttachment

Create a value of AwsEc2NetworkInterfaceAttachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:AwsEc2NetworkInterfaceAttachment', awsEc2NetworkInterfaceAttachment_instanceId - The ID of the instance.

$sel:status:AwsEc2NetworkInterfaceAttachment', awsEc2NetworkInterfaceAttachment_status - The attachment state.

Valid values: attaching | attached | detaching | detached

$sel:deleteOnTermination:AwsEc2NetworkInterfaceAttachment', awsEc2NetworkInterfaceAttachment_deleteOnTermination - Indicates whether the network interface is deleted when the instance is terminated.

$sel:attachmentId:AwsEc2NetworkInterfaceAttachment', awsEc2NetworkInterfaceAttachment_attachmentId - The identifier of the network interface attachment

$sel:instanceOwnerId:AwsEc2NetworkInterfaceAttachment', awsEc2NetworkInterfaceAttachment_instanceOwnerId - The Amazon Web Services account ID of the owner of the instance.

$sel:attachTime:AwsEc2NetworkInterfaceAttachment', awsEc2NetworkInterfaceAttachment_attachTime - Indicates when the attachment initiated.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:deviceIndex:AwsEc2NetworkInterfaceAttachment', awsEc2NetworkInterfaceAttachment_deviceIndex - The device index of the network interface attachment on the instance.

awsEc2NetworkInterfaceAttachment_instanceId :: Lens' AwsEc2NetworkInterfaceAttachment (Maybe Text)

The ID of the instance.

awsEc2NetworkInterfaceAttachment_status :: Lens' AwsEc2NetworkInterfaceAttachment (Maybe Text)

The attachment state.

Valid values: attaching | attached | detaching | detached

awsEc2NetworkInterfaceAttachment_deleteOnTermination :: Lens' AwsEc2NetworkInterfaceAttachment (Maybe Bool)

Indicates whether the network interface is deleted when the instance is terminated.

awsEc2NetworkInterfaceAttachment_attachmentId :: Lens' AwsEc2NetworkInterfaceAttachment (Maybe Text)

The identifier of the network interface attachment

awsEc2NetworkInterfaceAttachment_instanceOwnerId :: Lens' AwsEc2NetworkInterfaceAttachment (Maybe Text)

The Amazon Web Services account ID of the owner of the instance.

awsEc2NetworkInterfaceAttachment_attachTime :: Lens' AwsEc2NetworkInterfaceAttachment (Maybe Text)

Indicates when the attachment initiated.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsEc2NetworkInterfaceAttachment_deviceIndex :: Lens' AwsEc2NetworkInterfaceAttachment (Maybe Int)

The device index of the network interface attachment on the instance.