Network.AWS.SecurityHub.Types.AwsEc2NetworkInterfaceDetails

Description

Documentation

data AwsEc2NetworkInterfaceDetails

Details about the network interface

See: newAwsEc2NetworkInterfaceDetails smart constructor.

Constructors

AwsEc2NetworkInterfaceDetails'

Fields

privateIpAddresses :: Maybe [AwsEc2NetworkInterfacePrivateIpAddressDetail]

The private IPv4 addresses associated with the network interface.

publicDnsName :: Maybe Text

The public DNS name of the network interface.

securityGroups :: Maybe [AwsEc2NetworkInterfaceSecurityGroup]

Security groups for the network interface.

sourceDestCheck :: Maybe Bool

Indicates whether traffic to or from the instance is validated.

networkInterfaceId :: Maybe Text

The ID of the network interface.

attachment :: Maybe AwsEc2NetworkInterfaceAttachment

The network interface attachment.

publicIp :: Maybe Text

The address of the Elastic IP address bound to the network interface.

ipV6Addresses :: Maybe [AwsEc2NetworkInterfaceIpV6AddressDetail]

The IPv6 addresses associated with the network interface.

newAwsEc2NetworkInterfaceDetails :: AwsEc2NetworkInterfaceDetails

Create a value of AwsEc2NetworkInterfaceDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:privateIpAddresses:AwsEc2NetworkInterfaceDetails', awsEc2NetworkInterfaceDetails_privateIpAddresses - The private IPv4 addresses associated with the network interface.

$sel:publicDnsName:AwsEc2NetworkInterfaceDetails', awsEc2NetworkInterfaceDetails_publicDnsName - The public DNS name of the network interface.

$sel:securityGroups:AwsEc2NetworkInterfaceDetails', awsEc2NetworkInterfaceDetails_securityGroups - Security groups for the network interface.

$sel:sourceDestCheck:AwsEc2NetworkInterfaceDetails', awsEc2NetworkInterfaceDetails_sourceDestCheck - Indicates whether traffic to or from the instance is validated.

$sel:networkInterfaceId:AwsEc2NetworkInterfaceDetails', awsEc2NetworkInterfaceDetails_networkInterfaceId - The ID of the network interface.

$sel:attachment:AwsEc2NetworkInterfaceDetails', awsEc2NetworkInterfaceDetails_attachment - The network interface attachment.

$sel:publicIp:AwsEc2NetworkInterfaceDetails', awsEc2NetworkInterfaceDetails_publicIp - The address of the Elastic IP address bound to the network interface.

$sel:ipV6Addresses:AwsEc2NetworkInterfaceDetails', awsEc2NetworkInterfaceDetails_ipV6Addresses - The IPv6 addresses associated with the network interface.

awsEc2NetworkInterfaceDetails_privateIpAddresses :: Lens' AwsEc2NetworkInterfaceDetails (Maybe [AwsEc2NetworkInterfacePrivateIpAddressDetail])

The private IPv4 addresses associated with the network interface.

awsEc2NetworkInterfaceDetails_publicDnsName :: Lens' AwsEc2NetworkInterfaceDetails (Maybe Text)

The public DNS name of the network interface.

awsEc2NetworkInterfaceDetails_securityGroups :: Lens' AwsEc2NetworkInterfaceDetails (Maybe [AwsEc2NetworkInterfaceSecurityGroup])

Security groups for the network interface.

awsEc2NetworkInterfaceDetails_sourceDestCheck :: Lens' AwsEc2NetworkInterfaceDetails (Maybe Bool)

Indicates whether traffic to or from the instance is validated.

awsEc2NetworkInterfaceDetails_networkInterfaceId :: Lens' AwsEc2NetworkInterfaceDetails (Maybe Text)

The ID of the network interface.

awsEc2NetworkInterfaceDetails_attachment :: Lens' AwsEc2NetworkInterfaceDetails (Maybe AwsEc2NetworkInterfaceAttachment)

The network interface attachment.

awsEc2NetworkInterfaceDetails_publicIp :: Lens' AwsEc2NetworkInterfaceDetails (Maybe Text)

The address of the Elastic IP address bound to the network interface.

awsEc2NetworkInterfaceDetails_ipV6Addresses :: Lens' AwsEc2NetworkInterfaceDetails (Maybe [AwsEc2NetworkInterfaceIpV6AddressDetail])

The IPv6 addresses associated with the network interface.