Network.AWS.SecurityHub.Types.AwsEc2VpcDetails

Description

Documentation

data AwsEc2VpcDetails

Details about an EC2 VPC.

See: newAwsEc2VpcDetails smart constructor.

Constructors

AwsEc2VpcDetails'

Fields

state :: Maybe Text

The current state of the VPC.

ipv6CidrBlockAssociationSet :: Maybe [Ipv6CidrBlockAssociation]

Information about the IPv6 CIDR blocks associated with the VPC.

cidrBlockAssociationSet :: Maybe [CidrBlockAssociation]

Information about the IPv4 CIDR blocks associated with the VPC.

dhcpOptionsId :: Maybe Text

The identifier of the set of Dynamic Host Configuration Protocol (DHCP) options that are associated with the VPC. If the default options are associated with the VPC, then this is default.

newAwsEc2VpcDetails :: AwsEc2VpcDetails

Create a value of AwsEc2VpcDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:AwsEc2VpcDetails', awsEc2VpcDetails_state - The current state of the VPC.

$sel:ipv6CidrBlockAssociationSet:AwsEc2VpcDetails', awsEc2VpcDetails_ipv6CidrBlockAssociationSet - Information about the IPv6 CIDR blocks associated with the VPC.

$sel:cidrBlockAssociationSet:AwsEc2VpcDetails', awsEc2VpcDetails_cidrBlockAssociationSet - Information about the IPv4 CIDR blocks associated with the VPC.

$sel:dhcpOptionsId:AwsEc2VpcDetails', awsEc2VpcDetails_dhcpOptionsId - The identifier of the set of Dynamic Host Configuration Protocol (DHCP) options that are associated with the VPC. If the default options are associated with the VPC, then this is default.

awsEc2VpcDetails_state :: Lens' AwsEc2VpcDetails (Maybe Text)

The current state of the VPC.

awsEc2VpcDetails_ipv6CidrBlockAssociationSet :: Lens' AwsEc2VpcDetails (Maybe [Ipv6CidrBlockAssociation])

Information about the IPv6 CIDR blocks associated with the VPC.

awsEc2VpcDetails_cidrBlockAssociationSet :: Lens' AwsEc2VpcDetails (Maybe [CidrBlockAssociation])

Information about the IPv4 CIDR blocks associated with the VPC.

awsEc2VpcDetails_dhcpOptionsId :: Lens' AwsEc2VpcDetails (Maybe Text)

The identifier of the set of Dynamic Host Configuration Protocol (DHCP) options that are associated with the VPC. If the default options are associated with the VPC, then this is default.