Network.AWS.SecurityHub.Types.AwsEc2VpcEndpointServiceDetails

Description

Documentation

data AwsEc2VpcEndpointServiceDetails

Contains details about the service configuration for a VPC endpoint service.

See: newAwsEc2VpcEndpointServiceDetails smart constructor.

Constructors

AwsEc2VpcEndpointServiceDetails'

Fields

networkLoadBalancerArns :: Maybe [Text]

The ARNs of the Network Load Balancers for the service.

baseEndpointDnsNames :: Maybe [Text]

The DNS names for the service.

availabilityZones :: Maybe [Text]

The Availability Zones where the service is available.

gatewayLoadBalancerArns :: Maybe [Text]

The ARNs of the Gateway Load Balancers for the service.

managesVpcEndpoints :: Maybe Bool

Whether the service manages its VPC endpoints.

serviceName :: Maybe Text

The name of the service.

serviceState :: Maybe Text

The current state of the service.

serviceType :: Maybe [AwsEc2VpcEndpointServiceServiceTypeDetails]

The types for the service.

acceptanceRequired :: Maybe Bool

Whether requests from other Amazon Web Services accounts to create an endpoint to the service must first be accepted.

serviceId :: Maybe Text

The identifier of the service.

privateDnsName :: Maybe Text

The private DNS name for the service.

newAwsEc2VpcEndpointServiceDetails :: AwsEc2VpcEndpointServiceDetails

Create a value of AwsEc2VpcEndpointServiceDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkLoadBalancerArns:AwsEc2VpcEndpointServiceDetails', awsEc2VpcEndpointServiceDetails_networkLoadBalancerArns - The ARNs of the Network Load Balancers for the service.

$sel:baseEndpointDnsNames:AwsEc2VpcEndpointServiceDetails', awsEc2VpcEndpointServiceDetails_baseEndpointDnsNames - The DNS names for the service.

$sel:availabilityZones:AwsEc2VpcEndpointServiceDetails', awsEc2VpcEndpointServiceDetails_availabilityZones - The Availability Zones where the service is available.

$sel:gatewayLoadBalancerArns:AwsEc2VpcEndpointServiceDetails', awsEc2VpcEndpointServiceDetails_gatewayLoadBalancerArns - The ARNs of the Gateway Load Balancers for the service.

$sel:managesVpcEndpoints:AwsEc2VpcEndpointServiceDetails', awsEc2VpcEndpointServiceDetails_managesVpcEndpoints - Whether the service manages its VPC endpoints.

$sel:serviceName:AwsEc2VpcEndpointServiceDetails', awsEc2VpcEndpointServiceDetails_serviceName - The name of the service.

$sel:serviceState:AwsEc2VpcEndpointServiceDetails', awsEc2VpcEndpointServiceDetails_serviceState - The current state of the service.

$sel:serviceType:AwsEc2VpcEndpointServiceDetails', awsEc2VpcEndpointServiceDetails_serviceType - The types for the service.

$sel:acceptanceRequired:AwsEc2VpcEndpointServiceDetails', awsEc2VpcEndpointServiceDetails_acceptanceRequired - Whether requests from other Amazon Web Services accounts to create an endpoint to the service must first be accepted.

$sel:serviceId:AwsEc2VpcEndpointServiceDetails', awsEc2VpcEndpointServiceDetails_serviceId - The identifier of the service.

$sel:privateDnsName:AwsEc2VpcEndpointServiceDetails', awsEc2VpcEndpointServiceDetails_privateDnsName - The private DNS name for the service.

awsEc2VpcEndpointServiceDetails_networkLoadBalancerArns :: Lens' AwsEc2VpcEndpointServiceDetails (Maybe [Text])

The ARNs of the Network Load Balancers for the service.

awsEc2VpcEndpointServiceDetails_baseEndpointDnsNames :: Lens' AwsEc2VpcEndpointServiceDetails (Maybe [Text])

The DNS names for the service.

awsEc2VpcEndpointServiceDetails_availabilityZones :: Lens' AwsEc2VpcEndpointServiceDetails (Maybe [Text])

The Availability Zones where the service is available.

awsEc2VpcEndpointServiceDetails_gatewayLoadBalancerArns :: Lens' AwsEc2VpcEndpointServiceDetails (Maybe [Text])

The ARNs of the Gateway Load Balancers for the service.

awsEc2VpcEndpointServiceDetails_managesVpcEndpoints :: Lens' AwsEc2VpcEndpointServiceDetails (Maybe Bool)

Whether the service manages its VPC endpoints.

awsEc2VpcEndpointServiceDetails_serviceName :: Lens' AwsEc2VpcEndpointServiceDetails (Maybe Text)

The name of the service.

awsEc2VpcEndpointServiceDetails_serviceState :: Lens' AwsEc2VpcEndpointServiceDetails (Maybe Text)

The current state of the service.

awsEc2VpcEndpointServiceDetails_serviceType :: Lens' AwsEc2VpcEndpointServiceDetails (Maybe [AwsEc2VpcEndpointServiceServiceTypeDetails])

The types for the service.

awsEc2VpcEndpointServiceDetails_acceptanceRequired :: Lens' AwsEc2VpcEndpointServiceDetails (Maybe Bool)

Whether requests from other Amazon Web Services accounts to create an endpoint to the service must first be accepted.

awsEc2VpcEndpointServiceDetails_serviceId :: Lens' AwsEc2VpcEndpointServiceDetails (Maybe Text)

The identifier of the service.

awsEc2VpcEndpointServiceDetails_privateDnsName :: Lens' AwsEc2VpcEndpointServiceDetails (Maybe Text)

The private DNS name for the service.