Network.AWS.SecurityHub.Types.AwsEc2VpnConnectionOptionsTunnelOptionsDetails

Description

Documentation

data AwsEc2VpnConnectionOptionsTunnelOptionsDetails

The VPN tunnel options.

See: newAwsEc2VpnConnectionOptionsTunnelOptionsDetails smart constructor.

Constructors

AwsEc2VpnConnectionOptionsTunnelOptionsDetails'

Fields

outsideIpAddress :: Maybe Text

The external IP address of the VPN tunnel.

replayWindowSize :: Maybe Int

The number of packets in an IKE replay window.

rekeyFuzzPercentage :: Maybe Int

The percentage of the rekey window, which is determined by RekeyMarginTimeSeconds during which the rekey time is randomly selected.

phase1LifetimeSeconds :: Maybe Int

The lifetime for phase 1 of the IKE negotiation, in seconds.

ikeVersions :: Maybe [Text]

The Internet Key Exchange (IKE) versions that are permitted for the VPN tunnel.

phase2IntegrityAlgorithms :: Maybe [Text]

The permitted integrity algorithms for the VPN tunnel for phase 2 IKE negotiations.

phase2LifetimeSeconds :: Maybe Int

The lifetime for phase 2 of the IKE negotiation, in seconds.

phase1EncryptionAlgorithms :: Maybe [Text]

The permitted encryption algorithms for the VPN tunnel for phase 1 IKE negotiations.

phase1DhGroupNumbers :: Maybe [Int]

The permitted Diffie-Hellman group numbers for the VPN tunnel for phase 1 IKE negotiations.

phase1IntegrityAlgorithms :: Maybe [Text]

The permitted integrity algorithms for the VPN tunnel for phase 1 IKE negotiations.

rekeyMarginTimeSeconds :: Maybe Int

The margin time, in seconds, before the phase 2 lifetime expires, during which the Amazon Web Services side of the VPN connection performs an IKE rekey.

dpdTimeoutSeconds :: Maybe Int

The number of seconds after which a Dead Peer Detection (DPD) timeout occurs.

tunnelInsideCidr :: Maybe Text

The range of inside IPv4 addresses for the tunnel.

phase2EncryptionAlgorithms :: Maybe [Text]

The permitted encryption algorithms for the VPN tunnel for phase 2 IKE negotiations.

phase2DhGroupNumbers :: Maybe [Int]

The permitted Diffie-Hellman group numbers for the VPN tunnel for phase 2 IKE negotiations.

preSharedKey :: Maybe Text

The preshared key to establish initial authentication between the virtual private gateway and the customer gateway.

newAwsEc2VpnConnectionOptionsTunnelOptionsDetails :: AwsEc2VpnConnectionOptionsTunnelOptionsDetails

Create a value of AwsEc2VpnConnectionOptionsTunnelOptionsDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outsideIpAddress:AwsEc2VpnConnectionOptionsTunnelOptionsDetails', awsEc2VpnConnectionOptionsTunnelOptionsDetails_outsideIpAddress - The external IP address of the VPN tunnel.

$sel:replayWindowSize:AwsEc2VpnConnectionOptionsTunnelOptionsDetails', awsEc2VpnConnectionOptionsTunnelOptionsDetails_replayWindowSize - The number of packets in an IKE replay window.

$sel:rekeyFuzzPercentage:AwsEc2VpnConnectionOptionsTunnelOptionsDetails', awsEc2VpnConnectionOptionsTunnelOptionsDetails_rekeyFuzzPercentage - The percentage of the rekey window, which is determined by RekeyMarginTimeSeconds during which the rekey time is randomly selected.

$sel:phase1LifetimeSeconds:AwsEc2VpnConnectionOptionsTunnelOptionsDetails', awsEc2VpnConnectionOptionsTunnelOptionsDetails_phase1LifetimeSeconds - The lifetime for phase 1 of the IKE negotiation, in seconds.

$sel:ikeVersions:AwsEc2VpnConnectionOptionsTunnelOptionsDetails', awsEc2VpnConnectionOptionsTunnelOptionsDetails_ikeVersions - The Internet Key Exchange (IKE) versions that are permitted for the VPN tunnel.

$sel:phase2IntegrityAlgorithms:AwsEc2VpnConnectionOptionsTunnelOptionsDetails', awsEc2VpnConnectionOptionsTunnelOptionsDetails_phase2IntegrityAlgorithms - The permitted integrity algorithms for the VPN tunnel for phase 2 IKE negotiations.

$sel:phase2LifetimeSeconds:AwsEc2VpnConnectionOptionsTunnelOptionsDetails', awsEc2VpnConnectionOptionsTunnelOptionsDetails_phase2LifetimeSeconds - The lifetime for phase 2 of the IKE negotiation, in seconds.

$sel:phase1EncryptionAlgorithms:AwsEc2VpnConnectionOptionsTunnelOptionsDetails', awsEc2VpnConnectionOptionsTunnelOptionsDetails_phase1EncryptionAlgorithms - The permitted encryption algorithms for the VPN tunnel for phase 1 IKE negotiations.

$sel:phase1DhGroupNumbers:AwsEc2VpnConnectionOptionsTunnelOptionsDetails', awsEc2VpnConnectionOptionsTunnelOptionsDetails_phase1DhGroupNumbers - The permitted Diffie-Hellman group numbers for the VPN tunnel for phase 1 IKE negotiations.

$sel:phase1IntegrityAlgorithms:AwsEc2VpnConnectionOptionsTunnelOptionsDetails', awsEc2VpnConnectionOptionsTunnelOptionsDetails_phase1IntegrityAlgorithms - The permitted integrity algorithms for the VPN tunnel for phase 1 IKE negotiations.

$sel:rekeyMarginTimeSeconds:AwsEc2VpnConnectionOptionsTunnelOptionsDetails', awsEc2VpnConnectionOptionsTunnelOptionsDetails_rekeyMarginTimeSeconds - The margin time, in seconds, before the phase 2 lifetime expires, during which the Amazon Web Services side of the VPN connection performs an IKE rekey.

$sel:dpdTimeoutSeconds:AwsEc2VpnConnectionOptionsTunnelOptionsDetails', awsEc2VpnConnectionOptionsTunnelOptionsDetails_dpdTimeoutSeconds - The number of seconds after which a Dead Peer Detection (DPD) timeout occurs.

$sel:tunnelInsideCidr:AwsEc2VpnConnectionOptionsTunnelOptionsDetails', awsEc2VpnConnectionOptionsTunnelOptionsDetails_tunnelInsideCidr - The range of inside IPv4 addresses for the tunnel.

$sel:phase2EncryptionAlgorithms:AwsEc2VpnConnectionOptionsTunnelOptionsDetails', awsEc2VpnConnectionOptionsTunnelOptionsDetails_phase2EncryptionAlgorithms - The permitted encryption algorithms for the VPN tunnel for phase 2 IKE negotiations.

$sel:phase2DhGroupNumbers:AwsEc2VpnConnectionOptionsTunnelOptionsDetails', awsEc2VpnConnectionOptionsTunnelOptionsDetails_phase2DhGroupNumbers - The permitted Diffie-Hellman group numbers for the VPN tunnel for phase 2 IKE negotiations.

$sel:preSharedKey:AwsEc2VpnConnectionOptionsTunnelOptionsDetails', awsEc2VpnConnectionOptionsTunnelOptionsDetails_preSharedKey - The preshared key to establish initial authentication between the virtual private gateway and the customer gateway.

awsEc2VpnConnectionOptionsTunnelOptionsDetails_outsideIpAddress :: Lens' AwsEc2VpnConnectionOptionsTunnelOptionsDetails (Maybe Text)

The external IP address of the VPN tunnel.

awsEc2VpnConnectionOptionsTunnelOptionsDetails_replayWindowSize :: Lens' AwsEc2VpnConnectionOptionsTunnelOptionsDetails (Maybe Int)

The number of packets in an IKE replay window.

awsEc2VpnConnectionOptionsTunnelOptionsDetails_rekeyFuzzPercentage :: Lens' AwsEc2VpnConnectionOptionsTunnelOptionsDetails (Maybe Int)

The percentage of the rekey window, which is determined by RekeyMarginTimeSeconds during which the rekey time is randomly selected.

awsEc2VpnConnectionOptionsTunnelOptionsDetails_phase1LifetimeSeconds :: Lens' AwsEc2VpnConnectionOptionsTunnelOptionsDetails (Maybe Int)

The lifetime for phase 1 of the IKE negotiation, in seconds.

awsEc2VpnConnectionOptionsTunnelOptionsDetails_ikeVersions :: Lens' AwsEc2VpnConnectionOptionsTunnelOptionsDetails (Maybe [Text])

The Internet Key Exchange (IKE) versions that are permitted for the VPN tunnel.

awsEc2VpnConnectionOptionsTunnelOptionsDetails_phase2IntegrityAlgorithms :: Lens' AwsEc2VpnConnectionOptionsTunnelOptionsDetails (Maybe [Text])

The permitted integrity algorithms for the VPN tunnel for phase 2 IKE negotiations.

awsEc2VpnConnectionOptionsTunnelOptionsDetails_phase2LifetimeSeconds :: Lens' AwsEc2VpnConnectionOptionsTunnelOptionsDetails (Maybe Int)

The lifetime for phase 2 of the IKE negotiation, in seconds.

awsEc2VpnConnectionOptionsTunnelOptionsDetails_phase1EncryptionAlgorithms :: Lens' AwsEc2VpnConnectionOptionsTunnelOptionsDetails (Maybe [Text])

The permitted encryption algorithms for the VPN tunnel for phase 1 IKE negotiations.

awsEc2VpnConnectionOptionsTunnelOptionsDetails_phase1DhGroupNumbers :: Lens' AwsEc2VpnConnectionOptionsTunnelOptionsDetails (Maybe [Int])

The permitted Diffie-Hellman group numbers for the VPN tunnel for phase 1 IKE negotiations.

awsEc2VpnConnectionOptionsTunnelOptionsDetails_phase1IntegrityAlgorithms :: Lens' AwsEc2VpnConnectionOptionsTunnelOptionsDetails (Maybe [Text])

The permitted integrity algorithms for the VPN tunnel for phase 1 IKE negotiations.

awsEc2VpnConnectionOptionsTunnelOptionsDetails_rekeyMarginTimeSeconds :: Lens' AwsEc2VpnConnectionOptionsTunnelOptionsDetails (Maybe Int)

The margin time, in seconds, before the phase 2 lifetime expires, during which the Amazon Web Services side of the VPN connection performs an IKE rekey.

awsEc2VpnConnectionOptionsTunnelOptionsDetails_dpdTimeoutSeconds :: Lens' AwsEc2VpnConnectionOptionsTunnelOptionsDetails (Maybe Int)

The number of seconds after which a Dead Peer Detection (DPD) timeout occurs.

awsEc2VpnConnectionOptionsTunnelOptionsDetails_tunnelInsideCidr :: Lens' AwsEc2VpnConnectionOptionsTunnelOptionsDetails (Maybe Text)

The range of inside IPv4 addresses for the tunnel.

awsEc2VpnConnectionOptionsTunnelOptionsDetails_phase2EncryptionAlgorithms :: Lens' AwsEc2VpnConnectionOptionsTunnelOptionsDetails (Maybe [Text])

The permitted encryption algorithms for the VPN tunnel for phase 2 IKE negotiations.

awsEc2VpnConnectionOptionsTunnelOptionsDetails_phase2DhGroupNumbers :: Lens' AwsEc2VpnConnectionOptionsTunnelOptionsDetails (Maybe [Int])

The permitted Diffie-Hellman group numbers for the VPN tunnel for phase 2 IKE negotiations.

awsEc2VpnConnectionOptionsTunnelOptionsDetails_preSharedKey :: Lens' AwsEc2VpnConnectionOptionsTunnelOptionsDetails (Maybe Text)

The preshared key to establish initial authentication between the virtual private gateway and the customer gateway.