Network.AWS.SecurityHub.Types.AwsEcrContainerImageDetails

Description

Documentation

data AwsEcrContainerImageDetails

Information about an Amazon ECR image.

See: newAwsEcrContainerImageDetails smart constructor.

Constructors

AwsEcrContainerImageDetails'

Fields

registryId :: Maybe Text

The Amazon Web Services account identifier that is associated with the registry that the image belongs to.

imageTags :: Maybe [Text]

The list of tags that are associated with the image.

imageDigest :: Maybe Text

The sha256 digest of the image manifest.

imagePublishedAt :: Maybe Text

The date and time when the image was pushed to the repository.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

architecture :: Maybe Text

The architecture of the image.

repositoryName :: Maybe Text

The name of the repository that the image belongs to.

newAwsEcrContainerImageDetails :: AwsEcrContainerImageDetails

Create a value of AwsEcrContainerImageDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:AwsEcrContainerImageDetails', awsEcrContainerImageDetails_registryId - The Amazon Web Services account identifier that is associated with the registry that the image belongs to.

$sel:imageTags:AwsEcrContainerImageDetails', awsEcrContainerImageDetails_imageTags - The list of tags that are associated with the image.

$sel:imageDigest:AwsEcrContainerImageDetails', awsEcrContainerImageDetails_imageDigest - The sha256 digest of the image manifest.

$sel:imagePublishedAt:AwsEcrContainerImageDetails', awsEcrContainerImageDetails_imagePublishedAt - The date and time when the image was pushed to the repository.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:architecture:AwsEcrContainerImageDetails', awsEcrContainerImageDetails_architecture - The architecture of the image.

$sel:repositoryName:AwsEcrContainerImageDetails', awsEcrContainerImageDetails_repositoryName - The name of the repository that the image belongs to.

awsEcrContainerImageDetails_registryId :: Lens' AwsEcrContainerImageDetails (Maybe Text)

The Amazon Web Services account identifier that is associated with the registry that the image belongs to.

awsEcrContainerImageDetails_imageTags :: Lens' AwsEcrContainerImageDetails (Maybe [Text])

The list of tags that are associated with the image.

awsEcrContainerImageDetails_imageDigest :: Lens' AwsEcrContainerImageDetails (Maybe Text)

The sha256 digest of the image manifest.

awsEcrContainerImageDetails_imagePublishedAt :: Lens' AwsEcrContainerImageDetails (Maybe Text)

The date and time when the image was pushed to the repository.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsEcrContainerImageDetails_architecture :: Lens' AwsEcrContainerImageDetails (Maybe Text)

The architecture of the image.

awsEcrContainerImageDetails_repositoryName :: Lens' AwsEcrContainerImageDetails (Maybe Text)

The name of the repository that the image belongs to.