Network.AWS.SecurityHub.Types.AwsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails

Description

Documentation

data AwsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails

For tasks that use the awsvpc networking mode, the VPC subnet and security group configuration.

See: newAwsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails smart constructor.

Constructors

AwsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails'

Fields

securityGroups :: Maybe [Text]

The IDs of the security groups associated with the task or service.

You can provide up to five security groups.

subnets :: Maybe [Text]

The IDs of the subnets associated with the task or service.

You can provide up to 16 subnets.

assignPublicIp :: Maybe Text

Whether the task's elastic network interface receives a public IP address. The default value is DISABLED.

Valid values: ENABLED | DISABLED

newAwsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails :: AwsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails

Create a value of AwsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroups:AwsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails', awsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails_securityGroups - The IDs of the security groups associated with the task or service.

You can provide up to five security groups.

$sel:subnets:AwsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails', awsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails_subnets - The IDs of the subnets associated with the task or service.

You can provide up to 16 subnets.

$sel:assignPublicIp:AwsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails', awsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails_assignPublicIp - Whether the task's elastic network interface receives a public IP address. The default value is DISABLED.

Valid values: ENABLED | DISABLED

awsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails_securityGroups :: Lens' AwsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails (Maybe [Text])

The IDs of the security groups associated with the task or service.

You can provide up to five security groups.

awsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails_subnets :: Lens' AwsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails (Maybe [Text])

The IDs of the subnets associated with the task or service.

You can provide up to 16 subnets.

awsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails_assignPublicIp :: Lens' AwsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails (Maybe Text)

Whether the task's elastic network interface receives a public IP address. The default value is DISABLED.

Valid values: ENABLED | DISABLED