Network.AWS.SecurityHub.Types.AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails

Description

Documentation

data AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails

The container health check command and associated configuration parameters for the container.

See: newAwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails smart constructor.

Constructors

AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails'

Fields

command :: Maybe [Text]

The command that the container runs to determine whether it is healthy.

startPeriod :: Maybe Int

The optional grace period in seconds that allows containers time to bootstrap before failed health checks count towards the maximum number of retries.

retries :: Maybe Int

The number of times to retry a failed health check before the container is considered unhealthy. The default value is 3.

interval :: Maybe Int

The time period in seconds between each health check execution. The default value is 30 seconds.

timeout :: Maybe Int

The time period in seconds to wait for a health check to succeed before it is considered a failure. The default value is 5.

newAwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails :: AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails

Create a value of AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:command:AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails', awsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails_command - The command that the container runs to determine whether it is healthy.

$sel:startPeriod:AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails', awsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails_startPeriod - The optional grace period in seconds that allows containers time to bootstrap before failed health checks count towards the maximum number of retries.

$sel:retries:AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails', awsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails_retries - The number of times to retry a failed health check before the container is considered unhealthy. The default value is 3.

$sel:interval:AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails', awsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails_interval - The time period in seconds between each health check execution. The default value is 30 seconds.

$sel:timeout:AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails', awsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails_timeout - The time period in seconds to wait for a health check to succeed before it is considered a failure. The default value is 5.

awsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails_command :: Lens' AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails (Maybe [Text])

The command that the container runs to determine whether it is healthy.

awsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails_startPeriod :: Lens' AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails (Maybe Int)

The optional grace period in seconds that allows containers time to bootstrap before failed health checks count towards the maximum number of retries.

awsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails_retries :: Lens' AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails (Maybe Int)

The number of times to retry a failed health check before the container is considered unhealthy. The default value is 3.

awsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails_interval :: Lens' AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails (Maybe Int)

The time period in seconds between each health check execution. The default value is 30 seconds.

awsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails_timeout :: Lens' AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails (Maybe Int)

The time period in seconds to wait for a health check to succeed before it is considered a failure. The default value is 5.