Network.AWS.SecurityHub.Types.AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails

Description

Documentation

data AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails

The log configuration specification for the container.

See: newAwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails smart constructor.

Constructors

AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails'

Fields

logDriver :: Maybe Text

The log driver to use for the container.

options :: Maybe (HashMap Text Text)

The configuration options to send to the log driver. Requires version 1.19 of the Docker Remote API or greater on your container instance.

secretOptions :: Maybe [AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationSecretOptionsDetails]

The secrets to pass to the log configuration.

newAwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails :: AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails

Create a value of AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logDriver:AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails', awsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails_logDriver - The log driver to use for the container.

$sel:options:AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails', awsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails_options - The configuration options to send to the log driver. Requires version 1.19 of the Docker Remote API or greater on your container instance.

$sel:secretOptions:AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails', awsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails_secretOptions - The secrets to pass to the log configuration.

awsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails_logDriver :: Lens' AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails (Maybe Text)

The log driver to use for the container.

awsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails_options :: Lens' AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails (Maybe (HashMap Text Text))

The configuration options to send to the log driver. Requires version 1.19 of the Docker Remote API or greater on your container instance.

awsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails_secretOptions :: Lens' AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails (Maybe [AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationSecretOptionsDetails])

The secrets to pass to the log configuration.