Network.AWS.SecurityHub.Types.AwsEcsTaskDefinitionDetails

Description

Documentation

data AwsEcsTaskDefinitionDetails

details about a task definition. A task definition describes the container and volume definitions of an Amazon Elastic Container Service task.

See: newAwsEcsTaskDefinitionDetails smart constructor.

Constructors

AwsEcsTaskDefinitionDetails'

Fields

inferenceAccelerators :: Maybe [AwsEcsTaskDefinitionInferenceAcceleratorsDetails]

The Elastic Inference accelerators to use for the containers in the task.

executionRoleArn :: Maybe Text

The ARN of the task execution role that grants the container agent permission to make API calls on behalf of the container user.

pidMode :: Maybe Text

The process namespace to use for the containers in the task.

family :: Maybe Text

The name of a family that this task definition is registered to.

requiresCompatibilities :: Maybe [Text]

The task launch types that the task definition was validated against.

containerDefinitions :: Maybe [AwsEcsTaskDefinitionContainerDefinitionsDetails]

The container definitions that describe the containers that make up the task.

memory :: Maybe Text

The amount (in MiB) of memory used by the task.

ipcMode :: Maybe Text

The IPC resource namespace to use for the containers in the task.

taskRoleArn :: Maybe Text

The short name or ARN of the IAM role that grants containers in the task permission to call Amazon Web Services API operations on your behalf.

proxyConfiguration :: Maybe AwsEcsTaskDefinitionProxyConfigurationDetails

The configuration details for the App Mesh proxy.

placementConstraints :: Maybe [AwsEcsTaskDefinitionPlacementConstraintsDetails]

The placement constraint objects to use for tasks.

networkMode :: Maybe Text

The Docker networking mode to use for the containers in the task.

volumes :: Maybe [AwsEcsTaskDefinitionVolumesDetails]

The data volume definitions for the task.

cpu :: Maybe Text

The number of CPU units used by the task.

newAwsEcsTaskDefinitionDetails :: AwsEcsTaskDefinitionDetails

Create a value of AwsEcsTaskDefinitionDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inferenceAccelerators:AwsEcsTaskDefinitionDetails', awsEcsTaskDefinitionDetails_inferenceAccelerators - The Elastic Inference accelerators to use for the containers in the task.

$sel:executionRoleArn:AwsEcsTaskDefinitionDetails', awsEcsTaskDefinitionDetails_executionRoleArn - The ARN of the task execution role that grants the container agent permission to make API calls on behalf of the container user.

$sel:pidMode:AwsEcsTaskDefinitionDetails', awsEcsTaskDefinitionDetails_pidMode - The process namespace to use for the containers in the task.

$sel:family:AwsEcsTaskDefinitionDetails', awsEcsTaskDefinitionDetails_family - The name of a family that this task definition is registered to.

$sel:requiresCompatibilities:AwsEcsTaskDefinitionDetails', awsEcsTaskDefinitionDetails_requiresCompatibilities - The task launch types that the task definition was validated against.

$sel:containerDefinitions:AwsEcsTaskDefinitionDetails', awsEcsTaskDefinitionDetails_containerDefinitions - The container definitions that describe the containers that make up the task.

$sel:memory:AwsEcsTaskDefinitionDetails', awsEcsTaskDefinitionDetails_memory - The amount (in MiB) of memory used by the task.

$sel:ipcMode:AwsEcsTaskDefinitionDetails', awsEcsTaskDefinitionDetails_ipcMode - The IPC resource namespace to use for the containers in the task.

$sel:taskRoleArn:AwsEcsTaskDefinitionDetails', awsEcsTaskDefinitionDetails_taskRoleArn - The short name or ARN of the IAM role that grants containers in the task permission to call Amazon Web Services API operations on your behalf.

$sel:proxyConfiguration:AwsEcsTaskDefinitionDetails', awsEcsTaskDefinitionDetails_proxyConfiguration - The configuration details for the App Mesh proxy.

$sel:placementConstraints:AwsEcsTaskDefinitionDetails', awsEcsTaskDefinitionDetails_placementConstraints - The placement constraint objects to use for tasks.

$sel:networkMode:AwsEcsTaskDefinitionDetails', awsEcsTaskDefinitionDetails_networkMode - The Docker networking mode to use for the containers in the task.

$sel:volumes:AwsEcsTaskDefinitionDetails', awsEcsTaskDefinitionDetails_volumes - The data volume definitions for the task.

$sel:cpu:AwsEcsTaskDefinitionDetails', awsEcsTaskDefinitionDetails_cpu - The number of CPU units used by the task.

awsEcsTaskDefinitionDetails_inferenceAccelerators :: Lens' AwsEcsTaskDefinitionDetails (Maybe [AwsEcsTaskDefinitionInferenceAcceleratorsDetails])

The Elastic Inference accelerators to use for the containers in the task.

awsEcsTaskDefinitionDetails_executionRoleArn :: Lens' AwsEcsTaskDefinitionDetails (Maybe Text)

The ARN of the task execution role that grants the container agent permission to make API calls on behalf of the container user.

awsEcsTaskDefinitionDetails_pidMode :: Lens' AwsEcsTaskDefinitionDetails (Maybe Text)

The process namespace to use for the containers in the task.

awsEcsTaskDefinitionDetails_family :: Lens' AwsEcsTaskDefinitionDetails (Maybe Text)

The name of a family that this task definition is registered to.

awsEcsTaskDefinitionDetails_requiresCompatibilities :: Lens' AwsEcsTaskDefinitionDetails (Maybe [Text])

The task launch types that the task definition was validated against.

awsEcsTaskDefinitionDetails_containerDefinitions :: Lens' AwsEcsTaskDefinitionDetails (Maybe [AwsEcsTaskDefinitionContainerDefinitionsDetails])

The container definitions that describe the containers that make up the task.

awsEcsTaskDefinitionDetails_memory :: Lens' AwsEcsTaskDefinitionDetails (Maybe Text)

The amount (in MiB) of memory used by the task.

awsEcsTaskDefinitionDetails_ipcMode :: Lens' AwsEcsTaskDefinitionDetails (Maybe Text)

The IPC resource namespace to use for the containers in the task.

awsEcsTaskDefinitionDetails_taskRoleArn :: Lens' AwsEcsTaskDefinitionDetails (Maybe Text)

The short name or ARN of the IAM role that grants containers in the task permission to call Amazon Web Services API operations on your behalf.

awsEcsTaskDefinitionDetails_proxyConfiguration :: Lens' AwsEcsTaskDefinitionDetails (Maybe AwsEcsTaskDefinitionProxyConfigurationDetails)

The configuration details for the App Mesh proxy.

awsEcsTaskDefinitionDetails_placementConstraints :: Lens' AwsEcsTaskDefinitionDetails (Maybe [AwsEcsTaskDefinitionPlacementConstraintsDetails])

The placement constraint objects to use for tasks.

awsEcsTaskDefinitionDetails_networkMode :: Lens' AwsEcsTaskDefinitionDetails (Maybe Text)

The Docker networking mode to use for the containers in the task.

awsEcsTaskDefinitionDetails_volumes :: Lens' AwsEcsTaskDefinitionDetails (Maybe [AwsEcsTaskDefinitionVolumesDetails])

The data volume definitions for the task.

awsEcsTaskDefinitionDetails_cpu :: Lens' AwsEcsTaskDefinitionDetails (Maybe Text)

The number of CPU units used by the task.