Network.AWS.SecurityHub.Types.AwsEcsTaskDefinitionProxyConfigurationDetails

Description

Documentation

data AwsEcsTaskDefinitionProxyConfigurationDetails

The configuration details for the App Mesh proxy.

See: newAwsEcsTaskDefinitionProxyConfigurationDetails smart constructor.

Constructors

AwsEcsTaskDefinitionProxyConfigurationDetails'

Fields

proxyConfigurationProperties :: Maybe [AwsEcsTaskDefinitionProxyConfigurationProxyConfigurationPropertiesDetails]

The set of network configuration parameters to provide to the Container Network Interface (CNI) plugin, specified as key-value pairs.

containerName :: Maybe Text

The name of the container that will serve as the App Mesh proxy.

type' :: Maybe Text

The proxy type.

newAwsEcsTaskDefinitionProxyConfigurationDetails :: AwsEcsTaskDefinitionProxyConfigurationDetails

Create a value of AwsEcsTaskDefinitionProxyConfigurationDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:proxyConfigurationProperties:AwsEcsTaskDefinitionProxyConfigurationDetails', awsEcsTaskDefinitionProxyConfigurationDetails_proxyConfigurationProperties - The set of network configuration parameters to provide to the Container Network Interface (CNI) plugin, specified as key-value pairs.

$sel:containerName:AwsEcsTaskDefinitionProxyConfigurationDetails', awsEcsTaskDefinitionProxyConfigurationDetails_containerName - The name of the container that will serve as the App Mesh proxy.

$sel:type':AwsEcsTaskDefinitionProxyConfigurationDetails', awsEcsTaskDefinitionProxyConfigurationDetails_type - The proxy type.

awsEcsTaskDefinitionProxyConfigurationDetails_proxyConfigurationProperties :: Lens' AwsEcsTaskDefinitionProxyConfigurationDetails (Maybe [AwsEcsTaskDefinitionProxyConfigurationProxyConfigurationPropertiesDetails])

The set of network configuration parameters to provide to the Container Network Interface (CNI) plugin, specified as key-value pairs.

awsEcsTaskDefinitionProxyConfigurationDetails_containerName :: Lens' AwsEcsTaskDefinitionProxyConfigurationDetails (Maybe Text)

The name of the container that will serve as the App Mesh proxy.

awsEcsTaskDefinitionProxyConfigurationDetails_type :: Lens' AwsEcsTaskDefinitionProxyConfigurationDetails (Maybe Text)

The proxy type.