Network.AWS.SecurityHub.Types.AwsEcsTaskDefinitionVolumesDetails

Description

Documentation

data AwsEcsTaskDefinitionVolumesDetails

A data volume to mount from another container.

See: newAwsEcsTaskDefinitionVolumesDetails smart constructor.

Constructors

AwsEcsTaskDefinitionVolumesDetails'

Fields

dockerVolumeConfiguration :: Maybe AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails

Information about a Docker volume.

name :: Maybe Text

The name of the data volume.

efsVolumeConfiguration :: Maybe AwsEcsTaskDefinitionVolumesEfsVolumeConfigurationDetails

Information about the Amazon Elastic File System file system that is used for task storage.

host :: Maybe AwsEcsTaskDefinitionVolumesHostDetails

Information about a bind mount host volume.

newAwsEcsTaskDefinitionVolumesDetails :: AwsEcsTaskDefinitionVolumesDetails

Create a value of AwsEcsTaskDefinitionVolumesDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dockerVolumeConfiguration:AwsEcsTaskDefinitionVolumesDetails', awsEcsTaskDefinitionVolumesDetails_dockerVolumeConfiguration - Information about a Docker volume.

$sel:name:AwsEcsTaskDefinitionVolumesDetails', awsEcsTaskDefinitionVolumesDetails_name - The name of the data volume.

$sel:efsVolumeConfiguration:AwsEcsTaskDefinitionVolumesDetails', awsEcsTaskDefinitionVolumesDetails_efsVolumeConfiguration - Information about the Amazon Elastic File System file system that is used for task storage.

$sel:host:AwsEcsTaskDefinitionVolumesDetails', awsEcsTaskDefinitionVolumesDetails_host - Information about a bind mount host volume.

awsEcsTaskDefinitionVolumesDetails_dockerVolumeConfiguration :: Lens' AwsEcsTaskDefinitionVolumesDetails (Maybe AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails)

Information about a Docker volume.

awsEcsTaskDefinitionVolumesDetails_name :: Lens' AwsEcsTaskDefinitionVolumesDetails (Maybe Text)

The name of the data volume.

awsEcsTaskDefinitionVolumesDetails_efsVolumeConfiguration :: Lens' AwsEcsTaskDefinitionVolumesDetails (Maybe AwsEcsTaskDefinitionVolumesEfsVolumeConfigurationDetails)

Information about the Amazon Elastic File System file system that is used for task storage.

awsEcsTaskDefinitionVolumesDetails_host :: Lens' AwsEcsTaskDefinitionVolumesDetails (Maybe AwsEcsTaskDefinitionVolumesHostDetails)

Information about a bind mount host volume.