Network.AWS.SecurityHub.Types.AwsElasticBeanstalkEnvironmentOptionSetting

Description

Documentation

data AwsElasticBeanstalkEnvironmentOptionSetting

A configuration option setting for the environment.

See: newAwsElasticBeanstalkEnvironmentOptionSetting smart constructor.

Constructors

AwsElasticBeanstalkEnvironmentOptionSetting'

Fields

optionName :: Maybe Text

The name of the option.

resourceName :: Maybe Text

The name of the resource.

namespace :: Maybe Text

The type of resource that the configuration option is associated with.

value :: Maybe Text

The value of the configuration setting.

newAwsElasticBeanstalkEnvironmentOptionSetting :: AwsElasticBeanstalkEnvironmentOptionSetting

Create a value of AwsElasticBeanstalkEnvironmentOptionSetting with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:optionName:AwsElasticBeanstalkEnvironmentOptionSetting', awsElasticBeanstalkEnvironmentOptionSetting_optionName - The name of the option.

$sel:resourceName:AwsElasticBeanstalkEnvironmentOptionSetting', awsElasticBeanstalkEnvironmentOptionSetting_resourceName - The name of the resource.

$sel:namespace:AwsElasticBeanstalkEnvironmentOptionSetting', awsElasticBeanstalkEnvironmentOptionSetting_namespace - The type of resource that the configuration option is associated with.

$sel:value:AwsElasticBeanstalkEnvironmentOptionSetting', awsElasticBeanstalkEnvironmentOptionSetting_value - The value of the configuration setting.

awsElasticBeanstalkEnvironmentOptionSetting_optionName :: Lens' AwsElasticBeanstalkEnvironmentOptionSetting (Maybe Text)

The name of the option.

awsElasticBeanstalkEnvironmentOptionSetting_resourceName :: Lens' AwsElasticBeanstalkEnvironmentOptionSetting (Maybe Text)

The name of the resource.

awsElasticBeanstalkEnvironmentOptionSetting_namespace :: Lens' AwsElasticBeanstalkEnvironmentOptionSetting (Maybe Text)

The type of resource that the configuration option is associated with.

awsElasticBeanstalkEnvironmentOptionSetting_value :: Lens' AwsElasticBeanstalkEnvironmentOptionSetting (Maybe Text)

The value of the configuration setting.