Network.AWS.SecurityHub.Types.AwsElasticsearchDomainDetails

Description

Documentation

data AwsElasticsearchDomainDetails

Information about an Elasticsearch domain.

See: newAwsElasticsearchDomainDetails smart constructor.

Constructors

AwsElasticsearchDomainDetails'

Fields

nodeToNodeEncryptionOptions :: Maybe AwsElasticsearchDomainNodeToNodeEncryptionOptions

Details about the configuration for node-to-node encryption.

accessPolicies :: Maybe Text

IAM policy document specifying the access policies for the new Elasticsearch domain.

serviceSoftwareOptions :: Maybe AwsElasticsearchDomainServiceSoftwareOptions

Information about the status of a domain relative to the latest service software.

logPublishingOptions :: Maybe AwsElasticsearchDomainLogPublishingOptions

Configures the CloudWatch Logs to publish for the Elasticsearch domain.

elasticsearchClusterConfig :: Maybe AwsElasticsearchDomainElasticsearchClusterConfigDetails

Information about an OpenSearch cluster configuration.

domainName :: Maybe Text

Name of an Elasticsearch domain.

Domain names are unique across all domains owned by the same account within an Amazon Web Services Region.

Domain names must start with a lowercase letter and must be between 3 and 28 characters.

Valid characters are a-z (lowercase only), 0-9, and – (hyphen).

encryptionAtRestOptions :: Maybe AwsElasticsearchDomainEncryptionAtRestOptions

Details about the configuration for encryption at rest.

vPCOptions :: Maybe AwsElasticsearchDomainVPCOptions

Information that OpenSearch derives based on VPCOptions for the domain.

domainId :: Maybe Text

Unique identifier for an Elasticsearch domain.

endpoints :: Maybe (HashMap Text Text)

The key-value pair that exists if the Elasticsearch domain uses VPC endpoints.

domainEndpointOptions :: Maybe AwsElasticsearchDomainDomainEndpointOptions

Additional options for the domain endpoint.

endpoint :: Maybe Text

Domain-specific endpoint used to submit index, search, and data upload requests to an Elasticsearch domain.

The endpoint is a service URL.

elasticsearchVersion :: Maybe Text

OpenSearch version.

newAwsElasticsearchDomainDetails :: AwsElasticsearchDomainDetails

Create a value of AwsElasticsearchDomainDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nodeToNodeEncryptionOptions:AwsElasticsearchDomainDetails', awsElasticsearchDomainDetails_nodeToNodeEncryptionOptions - Details about the configuration for node-to-node encryption.

$sel:accessPolicies:AwsElasticsearchDomainDetails', awsElasticsearchDomainDetails_accessPolicies - IAM policy document specifying the access policies for the new Elasticsearch domain.

$sel:serviceSoftwareOptions:AwsElasticsearchDomainDetails', awsElasticsearchDomainDetails_serviceSoftwareOptions - Information about the status of a domain relative to the latest service software.

$sel:logPublishingOptions:AwsElasticsearchDomainDetails', awsElasticsearchDomainDetails_logPublishingOptions - Configures the CloudWatch Logs to publish for the Elasticsearch domain.

$sel:elasticsearchClusterConfig:AwsElasticsearchDomainDetails', awsElasticsearchDomainDetails_elasticsearchClusterConfig - Information about an OpenSearch cluster configuration.

$sel:domainName:AwsElasticsearchDomainDetails', awsElasticsearchDomainDetails_domainName - Name of an Elasticsearch domain.

Domain names are unique across all domains owned by the same account within an Amazon Web Services Region.

Domain names must start with a lowercase letter and must be between 3 and 28 characters.

Valid characters are a-z (lowercase only), 0-9, and – (hyphen).

$sel:encryptionAtRestOptions:AwsElasticsearchDomainDetails', awsElasticsearchDomainDetails_encryptionAtRestOptions - Details about the configuration for encryption at rest.

$sel:vPCOptions:AwsElasticsearchDomainDetails', awsElasticsearchDomainDetails_vPCOptions - Information that OpenSearch derives based on VPCOptions for the domain.

$sel:domainId:AwsElasticsearchDomainDetails', awsElasticsearchDomainDetails_domainId - Unique identifier for an Elasticsearch domain.

$sel:endpoints:AwsElasticsearchDomainDetails', awsElasticsearchDomainDetails_endpoints - The key-value pair that exists if the Elasticsearch domain uses VPC endpoints.

$sel:domainEndpointOptions:AwsElasticsearchDomainDetails', awsElasticsearchDomainDetails_domainEndpointOptions - Additional options for the domain endpoint.

$sel:endpoint:AwsElasticsearchDomainDetails', awsElasticsearchDomainDetails_endpoint - Domain-specific endpoint used to submit index, search, and data upload requests to an Elasticsearch domain.

The endpoint is a service URL.

$sel:elasticsearchVersion:AwsElasticsearchDomainDetails', awsElasticsearchDomainDetails_elasticsearchVersion - OpenSearch version.

awsElasticsearchDomainDetails_nodeToNodeEncryptionOptions :: Lens' AwsElasticsearchDomainDetails (Maybe AwsElasticsearchDomainNodeToNodeEncryptionOptions)

Details about the configuration for node-to-node encryption.

awsElasticsearchDomainDetails_accessPolicies :: Lens' AwsElasticsearchDomainDetails (Maybe Text)

IAM policy document specifying the access policies for the new Elasticsearch domain.

awsElasticsearchDomainDetails_serviceSoftwareOptions :: Lens' AwsElasticsearchDomainDetails (Maybe AwsElasticsearchDomainServiceSoftwareOptions)

Information about the status of a domain relative to the latest service software.

awsElasticsearchDomainDetails_logPublishingOptions :: Lens' AwsElasticsearchDomainDetails (Maybe AwsElasticsearchDomainLogPublishingOptions)

Configures the CloudWatch Logs to publish for the Elasticsearch domain.

awsElasticsearchDomainDetails_elasticsearchClusterConfig :: Lens' AwsElasticsearchDomainDetails (Maybe AwsElasticsearchDomainElasticsearchClusterConfigDetails)

Information about an OpenSearch cluster configuration.

awsElasticsearchDomainDetails_domainName :: Lens' AwsElasticsearchDomainDetails (Maybe Text)

Name of an Elasticsearch domain.

Domain names are unique across all domains owned by the same account within an Amazon Web Services Region.

Domain names must start with a lowercase letter and must be between 3 and 28 characters.

Valid characters are a-z (lowercase only), 0-9, and – (hyphen).

awsElasticsearchDomainDetails_encryptionAtRestOptions :: Lens' AwsElasticsearchDomainDetails (Maybe AwsElasticsearchDomainEncryptionAtRestOptions)

Details about the configuration for encryption at rest.

awsElasticsearchDomainDetails_vPCOptions :: Lens' AwsElasticsearchDomainDetails (Maybe AwsElasticsearchDomainVPCOptions)

Information that OpenSearch derives based on VPCOptions for the domain.

awsElasticsearchDomainDetails_domainId :: Lens' AwsElasticsearchDomainDetails (Maybe Text)

Unique identifier for an Elasticsearch domain.

awsElasticsearchDomainDetails_endpoints :: Lens' AwsElasticsearchDomainDetails (Maybe (HashMap Text Text))

The key-value pair that exists if the Elasticsearch domain uses VPC endpoints.

awsElasticsearchDomainDetails_domainEndpointOptions :: Lens' AwsElasticsearchDomainDetails (Maybe AwsElasticsearchDomainDomainEndpointOptions)

Additional options for the domain endpoint.

awsElasticsearchDomainDetails_endpoint :: Lens' AwsElasticsearchDomainDetails (Maybe Text)

Domain-specific endpoint used to submit index, search, and data upload requests to an Elasticsearch domain.

The endpoint is a service URL.

awsElasticsearchDomainDetails_elasticsearchVersion :: Lens' AwsElasticsearchDomainDetails (Maybe Text)

OpenSearch version.