Network.AWS.SecurityHub.Types.AwsElasticsearchDomainDomainEndpointOptions

Description

Documentation

data AwsElasticsearchDomainDomainEndpointOptions

Additional options for the domain endpoint, such as whether to require HTTPS for all traffic.

See: newAwsElasticsearchDomainDomainEndpointOptions smart constructor.

Constructors

AwsElasticsearchDomainDomainEndpointOptions'

Fields

enforceHTTPS :: Maybe Bool

Whether to require that all traffic to the domain arrive over HTTPS.

tLSSecurityPolicy :: Maybe Text

The TLS security policy to apply to the HTTPS endpoint of the OpenSearch domain.

Valid values:

Policy-Min-TLS-1-0-2019-07, which supports TLSv1.0 and higher

Policy-Min-TLS-1-2-2019-07, which only supports TLSv1.2

newAwsElasticsearchDomainDomainEndpointOptions :: AwsElasticsearchDomainDomainEndpointOptions

Create a value of AwsElasticsearchDomainDomainEndpointOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enforceHTTPS:AwsElasticsearchDomainDomainEndpointOptions', awsElasticsearchDomainDomainEndpointOptions_enforceHTTPS - Whether to require that all traffic to the domain arrive over HTTPS.

$sel:tLSSecurityPolicy:AwsElasticsearchDomainDomainEndpointOptions', awsElasticsearchDomainDomainEndpointOptions_tLSSecurityPolicy - The TLS security policy to apply to the HTTPS endpoint of the OpenSearch domain.

Valid values:

Policy-Min-TLS-1-0-2019-07, which supports TLSv1.0 and higher

Policy-Min-TLS-1-2-2019-07, which only supports TLSv1.2

awsElasticsearchDomainDomainEndpointOptions_enforceHTTPS :: Lens' AwsElasticsearchDomainDomainEndpointOptions (Maybe Bool)

Whether to require that all traffic to the domain arrive over HTTPS.

awsElasticsearchDomainDomainEndpointOptions_tLSSecurityPolicy :: Lens' AwsElasticsearchDomainDomainEndpointOptions (Maybe Text)

The TLS security policy to apply to the HTTPS endpoint of the OpenSearch domain.

Valid values:

Policy-Min-TLS-1-0-2019-07, which supports TLSv1.0 and higher

Policy-Min-TLS-1-2-2019-07, which only supports TLSv1.2