Network.AWS.SecurityHub.Types.AwsElasticsearchDomainVPCOptions

Description

Documentation

data AwsElasticsearchDomainVPCOptions

Information that OpenSearch derives based on VPCOptions for the domain.

See: newAwsElasticsearchDomainVPCOptions smart constructor.

Constructors

AwsElasticsearchDomainVPCOptions'

Fields

securityGroupIds :: Maybe [Text]

The list of security group IDs associated with the VPC endpoints for the domain.

subnetIds :: Maybe [Text]

A list of subnet IDs associated with the VPC endpoints for the domain.

vPCId :: Maybe Text

ID for the VPC.

availabilityZones :: Maybe [Text]

The list of Availability Zones associated with the VPC subnets.

newAwsElasticsearchDomainVPCOptions :: AwsElasticsearchDomainVPCOptions

Create a value of AwsElasticsearchDomainVPCOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:AwsElasticsearchDomainVPCOptions', awsElasticsearchDomainVPCOptions_securityGroupIds - The list of security group IDs associated with the VPC endpoints for the domain.

$sel:subnetIds:AwsElasticsearchDomainVPCOptions', awsElasticsearchDomainVPCOptions_subnetIds - A list of subnet IDs associated with the VPC endpoints for the domain.

$sel:vPCId:AwsElasticsearchDomainVPCOptions', awsElasticsearchDomainVPCOptions_vPCId - ID for the VPC.

$sel:availabilityZones:AwsElasticsearchDomainVPCOptions', awsElasticsearchDomainVPCOptions_availabilityZones - The list of Availability Zones associated with the VPC subnets.

awsElasticsearchDomainVPCOptions_securityGroupIds :: Lens' AwsElasticsearchDomainVPCOptions (Maybe [Text])

The list of security group IDs associated with the VPC endpoints for the domain.

awsElasticsearchDomainVPCOptions_subnetIds :: Lens' AwsElasticsearchDomainVPCOptions (Maybe [Text])

A list of subnet IDs associated with the VPC endpoints for the domain.

awsElasticsearchDomainVPCOptions_vPCId :: Lens' AwsElasticsearchDomainVPCOptions (Maybe Text)

ID for the VPC.

awsElasticsearchDomainVPCOptions_availabilityZones :: Lens' AwsElasticsearchDomainVPCOptions (Maybe [Text])

The list of Availability Zones associated with the VPC subnets.