Network.AWS.SecurityHub.Types.AwsElbLoadBalancerAccessLog

Description

Documentation

data AwsElbLoadBalancerAccessLog

Contains information about the access log configuration for the load balancer.

See: newAwsElbLoadBalancerAccessLog smart constructor.

Constructors

AwsElbLoadBalancerAccessLog'

Fields

emitInterval :: Maybe Int

The interval in minutes for publishing the access logs.

You can publish access logs either every 5 minutes or every 60 minutes.

enabled :: Maybe Bool

Indicates whether access logs are enabled for the load balancer.

s3BucketPrefix :: Maybe Text

The logical hierarchy that was created for the S3 bucket.

If a prefix is not provided, the log is placed at the root level of the bucket.

s3BucketName :: Maybe Text

The name of the S3 bucket where the access logs are stored.

newAwsElbLoadBalancerAccessLog :: AwsElbLoadBalancerAccessLog

Create a value of AwsElbLoadBalancerAccessLog with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:emitInterval:AwsElbLoadBalancerAccessLog', awsElbLoadBalancerAccessLog_emitInterval - The interval in minutes for publishing the access logs.

You can publish access logs either every 5 minutes or every 60 minutes.

$sel:enabled:AwsElbLoadBalancerAccessLog', awsElbLoadBalancerAccessLog_enabled - Indicates whether access logs are enabled for the load balancer.

$sel:s3BucketPrefix:AwsElbLoadBalancerAccessLog', awsElbLoadBalancerAccessLog_s3BucketPrefix - The logical hierarchy that was created for the S3 bucket.

If a prefix is not provided, the log is placed at the root level of the bucket.

$sel:s3BucketName:AwsElbLoadBalancerAccessLog', awsElbLoadBalancerAccessLog_s3BucketName - The name of the S3 bucket where the access logs are stored.

awsElbLoadBalancerAccessLog_emitInterval :: Lens' AwsElbLoadBalancerAccessLog (Maybe Int)

The interval in minutes for publishing the access logs.

You can publish access logs either every 5 minutes or every 60 minutes.

awsElbLoadBalancerAccessLog_enabled :: Lens' AwsElbLoadBalancerAccessLog (Maybe Bool)

Indicates whether access logs are enabled for the load balancer.

awsElbLoadBalancerAccessLog_s3BucketPrefix :: Lens' AwsElbLoadBalancerAccessLog (Maybe Text)

The logical hierarchy that was created for the S3 bucket.

If a prefix is not provided, the log is placed at the root level of the bucket.

awsElbLoadBalancerAccessLog_s3BucketName :: Lens' AwsElbLoadBalancerAccessLog (Maybe Text)

The name of the S3 bucket where the access logs are stored.