Network.AWS.SecurityHub.Types.AwsElbLoadBalancerAttributes

Description

Documentation

data AwsElbLoadBalancerAttributes

Contains attributes for the load balancer.

See: newAwsElbLoadBalancerAttributes smart constructor.

Constructors

AwsElbLoadBalancerAttributes'

Fields

crossZoneLoadBalancing :: Maybe AwsElbLoadBalancerCrossZoneLoadBalancing

Cross-zone load balancing settings for the load balancer.

If cross-zone load balancing is enabled, the load balancer routes the request traffic evenly across all instances regardless of the Availability Zones.

accessLog :: Maybe AwsElbLoadBalancerAccessLog

Information about the access log configuration for the load balancer.

If the access log is enabled, the load balancer captures detailed information about all requests. It delivers the information to a specified S3 bucket.

connectionSettings :: Maybe AwsElbLoadBalancerConnectionSettings

Connection settings for the load balancer.

If an idle timeout is configured, the load balancer allows connections to remain idle for the specified duration. When a connection is idle, no data is sent over the connection.

connectionDraining :: Maybe AwsElbLoadBalancerConnectionDraining

Information about the connection draining configuration for the load balancer.

If connection draining is enabled, the load balancer allows existing requests to complete before it shifts traffic away from a deregistered or unhealthy instance.

newAwsElbLoadBalancerAttributes :: AwsElbLoadBalancerAttributes

Create a value of AwsElbLoadBalancerAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:crossZoneLoadBalancing:AwsElbLoadBalancerAttributes', awsElbLoadBalancerAttributes_crossZoneLoadBalancing - Cross-zone load balancing settings for the load balancer.

If cross-zone load balancing is enabled, the load balancer routes the request traffic evenly across all instances regardless of the Availability Zones.

$sel:accessLog:AwsElbLoadBalancerAttributes', awsElbLoadBalancerAttributes_accessLog - Information about the access log configuration for the load balancer.

If the access log is enabled, the load balancer captures detailed information about all requests. It delivers the information to a specified S3 bucket.

$sel:connectionSettings:AwsElbLoadBalancerAttributes', awsElbLoadBalancerAttributes_connectionSettings - Connection settings for the load balancer.

If an idle timeout is configured, the load balancer allows connections to remain idle for the specified duration. When a connection is idle, no data is sent over the connection.

$sel:connectionDraining:AwsElbLoadBalancerAttributes', awsElbLoadBalancerAttributes_connectionDraining - Information about the connection draining configuration for the load balancer.

If connection draining is enabled, the load balancer allows existing requests to complete before it shifts traffic away from a deregistered or unhealthy instance.

awsElbLoadBalancerAttributes_crossZoneLoadBalancing :: Lens' AwsElbLoadBalancerAttributes (Maybe AwsElbLoadBalancerCrossZoneLoadBalancing)

Cross-zone load balancing settings for the load balancer.

If cross-zone load balancing is enabled, the load balancer routes the request traffic evenly across all instances regardless of the Availability Zones.

awsElbLoadBalancerAttributes_accessLog :: Lens' AwsElbLoadBalancerAttributes (Maybe AwsElbLoadBalancerAccessLog)

Information about the access log configuration for the load balancer.

If the access log is enabled, the load balancer captures detailed information about all requests. It delivers the information to a specified S3 bucket.

awsElbLoadBalancerAttributes_connectionSettings :: Lens' AwsElbLoadBalancerAttributes (Maybe AwsElbLoadBalancerConnectionSettings)

Connection settings for the load balancer.

If an idle timeout is configured, the load balancer allows connections to remain idle for the specified duration. When a connection is idle, no data is sent over the connection.

awsElbLoadBalancerAttributes_connectionDraining :: Lens' AwsElbLoadBalancerAttributes (Maybe AwsElbLoadBalancerConnectionDraining)

Information about the connection draining configuration for the load balancer.

If connection draining is enabled, the load balancer allows existing requests to complete before it shifts traffic away from a deregistered or unhealthy instance.