Network.AWS.SecurityHub.Types.AwsElbv2LoadBalancerDetails

Description

Documentation

data AwsElbv2LoadBalancerDetails

Information about a load balancer.

See: newAwsElbv2LoadBalancerDetails smart constructor.

Constructors

AwsElbv2LoadBalancerDetails'

Fields

state :: Maybe LoadBalancerState

The state of the load balancer.

securityGroups :: Maybe [Text]

The IDs of the security groups for the load balancer.

loadBalancerAttributes :: Maybe [AwsElbv2LoadBalancerAttribute]

Attributes of the load balancer.

createdTime :: Maybe Text

Indicates when the load balancer was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

vpcId :: Maybe Text

The ID of the VPC for the load balancer.

canonicalHostedZoneId :: Maybe Text

The ID of the Amazon Route 53 hosted zone associated with the load balancer.

availabilityZones :: Maybe [AvailabilityZone]

The Availability Zones for the load balancer.

ipAddressType :: Maybe Text

The type of IP addresses used by the subnets for your load balancer. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses).

scheme :: Maybe Text

The nodes of an Internet-facing load balancer have public IP addresses.

type' :: Maybe Text

The type of load balancer.

dNSName :: Maybe Text

The public DNS name of the load balancer.

newAwsElbv2LoadBalancerDetails :: AwsElbv2LoadBalancerDetails

Create a value of AwsElbv2LoadBalancerDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:AwsElbv2LoadBalancerDetails', awsElbv2LoadBalancerDetails_state - The state of the load balancer.

$sel:securityGroups:AwsElbv2LoadBalancerDetails', awsElbv2LoadBalancerDetails_securityGroups - The IDs of the security groups for the load balancer.

$sel:loadBalancerAttributes:AwsElbv2LoadBalancerDetails', awsElbv2LoadBalancerDetails_loadBalancerAttributes - Attributes of the load balancer.

$sel:createdTime:AwsElbv2LoadBalancerDetails', awsElbv2LoadBalancerDetails_createdTime - Indicates when the load balancer was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:vpcId:AwsElbv2LoadBalancerDetails', awsElbv2LoadBalancerDetails_vpcId - The ID of the VPC for the load balancer.

$sel:canonicalHostedZoneId:AwsElbv2LoadBalancerDetails', awsElbv2LoadBalancerDetails_canonicalHostedZoneId - The ID of the Amazon Route 53 hosted zone associated with the load balancer.

$sel:availabilityZones:AwsElbv2LoadBalancerDetails', awsElbv2LoadBalancerDetails_availabilityZones - The Availability Zones for the load balancer.

$sel:ipAddressType:AwsElbv2LoadBalancerDetails', awsElbv2LoadBalancerDetails_ipAddressType - The type of IP addresses used by the subnets for your load balancer. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses).

$sel:scheme:AwsElbv2LoadBalancerDetails', awsElbv2LoadBalancerDetails_scheme - The nodes of an Internet-facing load balancer have public IP addresses.

$sel:type':AwsElbv2LoadBalancerDetails', awsElbv2LoadBalancerDetails_type - The type of load balancer.

$sel:dNSName:AwsElbv2LoadBalancerDetails', awsElbv2LoadBalancerDetails_dNSName - The public DNS name of the load balancer.

awsElbv2LoadBalancerDetails_state :: Lens' AwsElbv2LoadBalancerDetails (Maybe LoadBalancerState)

The state of the load balancer.

awsElbv2LoadBalancerDetails_securityGroups :: Lens' AwsElbv2LoadBalancerDetails (Maybe [Text])

The IDs of the security groups for the load balancer.

awsElbv2LoadBalancerDetails_loadBalancerAttributes :: Lens' AwsElbv2LoadBalancerDetails (Maybe [AwsElbv2LoadBalancerAttribute])

Attributes of the load balancer.

awsElbv2LoadBalancerDetails_createdTime :: Lens' AwsElbv2LoadBalancerDetails (Maybe Text)

Indicates when the load balancer was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsElbv2LoadBalancerDetails_vpcId :: Lens' AwsElbv2LoadBalancerDetails (Maybe Text)

The ID of the VPC for the load balancer.

awsElbv2LoadBalancerDetails_canonicalHostedZoneId :: Lens' AwsElbv2LoadBalancerDetails (Maybe Text)

The ID of the Amazon Route 53 hosted zone associated with the load balancer.

awsElbv2LoadBalancerDetails_availabilityZones :: Lens' AwsElbv2LoadBalancerDetails (Maybe [AvailabilityZone])

The Availability Zones for the load balancer.

awsElbv2LoadBalancerDetails_ipAddressType :: Lens' AwsElbv2LoadBalancerDetails (Maybe Text)

The type of IP addresses used by the subnets for your load balancer. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses).

awsElbv2LoadBalancerDetails_scheme :: Lens' AwsElbv2LoadBalancerDetails (Maybe Text)

The nodes of an Internet-facing load balancer have public IP addresses.

awsElbv2LoadBalancerDetails_type :: Lens' AwsElbv2LoadBalancerDetails (Maybe Text)

The type of load balancer.

awsElbv2LoadBalancerDetails_dNSName :: Lens' AwsElbv2LoadBalancerDetails (Maybe Text)

The public DNS name of the load balancer.