Network.AWS.SecurityHub.Types.AwsIamGroupDetails

Description

Documentation

data AwsIamGroupDetails

Contains details about an IAM group.

See: newAwsIamGroupDetails smart constructor.

Constructors

AwsIamGroupDetails'

Fields

path :: Maybe Text

The path to the group.

createDate :: Maybe Text

Indicates when the IAM group was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

groupId :: Maybe Text

The identifier of the IAM group.

groupPolicyList :: Maybe [AwsIamGroupPolicy]

The list of inline policies that are embedded in the group.

groupName :: Maybe Text

The name of the IAM group.

attachedManagedPolicies :: Maybe [AwsIamAttachedManagedPolicy]

A list of the managed policies that are attached to the IAM group.

newAwsIamGroupDetails :: AwsIamGroupDetails

Create a value of AwsIamGroupDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:path:AwsIamGroupDetails', awsIamGroupDetails_path - The path to the group.

$sel:createDate:AwsIamGroupDetails', awsIamGroupDetails_createDate - Indicates when the IAM group was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:groupId:AwsIamGroupDetails', awsIamGroupDetails_groupId - The identifier of the IAM group.

$sel:groupPolicyList:AwsIamGroupDetails', awsIamGroupDetails_groupPolicyList - The list of inline policies that are embedded in the group.

$sel:groupName:AwsIamGroupDetails', awsIamGroupDetails_groupName - The name of the IAM group.

$sel:attachedManagedPolicies:AwsIamGroupDetails', awsIamGroupDetails_attachedManagedPolicies - A list of the managed policies that are attached to the IAM group.

awsIamGroupDetails_path :: Lens' AwsIamGroupDetails (Maybe Text)

The path to the group.

awsIamGroupDetails_createDate :: Lens' AwsIamGroupDetails (Maybe Text)

Indicates when the IAM group was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsIamGroupDetails_groupId :: Lens' AwsIamGroupDetails (Maybe Text)

The identifier of the IAM group.

awsIamGroupDetails_groupPolicyList :: Lens' AwsIamGroupDetails (Maybe [AwsIamGroupPolicy])

The list of inline policies that are embedded in the group.

awsIamGroupDetails_groupName :: Lens' AwsIamGroupDetails (Maybe Text)

The name of the IAM group.

awsIamGroupDetails_attachedManagedPolicies :: Lens' AwsIamGroupDetails (Maybe [AwsIamAttachedManagedPolicy])

A list of the managed policies that are attached to the IAM group.