Network.AWS.SecurityHub.Types.AwsLambdaFunctionDetails

Description

Documentation

data AwsLambdaFunctionDetails

Details about a function's configuration.

See: newAwsLambdaFunctionDetails smart constructor.

Constructors

AwsLambdaFunctionDetails'

Fields

memorySize :: Maybe Int

The memory that is allocated to the function.

runtime :: Maybe Text

The runtime environment for the Lambda function.

kmsKeyArn :: Maybe Text

The KMS key that is used to encrypt the function's environment variables. This key is only returned if you've configured a customer managed customer managed key.

environment :: Maybe AwsLambdaFunctionEnvironment

The function's environment variables.

deadLetterConfig :: Maybe AwsLambdaFunctionDeadLetterConfig

The function's dead letter queue.

role' :: Maybe Text

The function's execution role.

vpcConfig :: Maybe AwsLambdaFunctionVpcConfig

The function's networking configuration.

version :: Maybe Text

The version of the Lambda function.

functionName :: Maybe Text

The name of the function.

code :: Maybe AwsLambdaFunctionCode

An AwsLambdaFunctionCode object.

layers :: Maybe [AwsLambdaFunctionLayer]

The function's layers.

handler :: Maybe Text

The function that Lambda calls to begin executing your function.

timeout :: Maybe Int

The amount of time that Lambda allows a function to run before stopping it.

lastModified :: Maybe Text

Indicates when the function was last updated.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

codeSha256 :: Maybe Text

The SHA256 hash of the function's deployment package.

tracingConfig :: Maybe AwsLambdaFunctionTracingConfig

The function's X-Ray tracing configuration.

revisionId :: Maybe Text

The latest updated revision of the function or alias.

masterArn :: Maybe Text

For Lambda@Edge functions, the ARN of the master function.

newAwsLambdaFunctionDetails :: AwsLambdaFunctionDetails

Create a value of AwsLambdaFunctionDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memorySize:AwsLambdaFunctionDetails', awsLambdaFunctionDetails_memorySize - The memory that is allocated to the function.

$sel:runtime:AwsLambdaFunctionDetails', awsLambdaFunctionDetails_runtime - The runtime environment for the Lambda function.

$sel:kmsKeyArn:AwsLambdaFunctionDetails', awsLambdaFunctionDetails_kmsKeyArn - The KMS key that is used to encrypt the function's environment variables. This key is only returned if you've configured a customer managed customer managed key.

$sel:environment:AwsLambdaFunctionDetails', awsLambdaFunctionDetails_environment - The function's environment variables.

$sel:deadLetterConfig:AwsLambdaFunctionDetails', awsLambdaFunctionDetails_deadLetterConfig - The function's dead letter queue.

$sel:role':AwsLambdaFunctionDetails', awsLambdaFunctionDetails_role - The function's execution role.

$sel:vpcConfig:AwsLambdaFunctionDetails', awsLambdaFunctionDetails_vpcConfig - The function's networking configuration.

$sel:version:AwsLambdaFunctionDetails', awsLambdaFunctionDetails_version - The version of the Lambda function.

$sel:functionName:AwsLambdaFunctionDetails', awsLambdaFunctionDetails_functionName - The name of the function.

$sel:code:AwsLambdaFunctionDetails', awsLambdaFunctionDetails_code - An AwsLambdaFunctionCode object.

$sel:layers:AwsLambdaFunctionDetails', awsLambdaFunctionDetails_layers - The function's layers.

$sel:handler:AwsLambdaFunctionDetails', awsLambdaFunctionDetails_handler - The function that Lambda calls to begin executing your function.

$sel:timeout:AwsLambdaFunctionDetails', awsLambdaFunctionDetails_timeout - The amount of time that Lambda allows a function to run before stopping it.

$sel:lastModified:AwsLambdaFunctionDetails', awsLambdaFunctionDetails_lastModified - Indicates when the function was last updated.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:codeSha256:AwsLambdaFunctionDetails', awsLambdaFunctionDetails_codeSha256 - The SHA256 hash of the function's deployment package.

$sel:tracingConfig:AwsLambdaFunctionDetails', awsLambdaFunctionDetails_tracingConfig - The function's X-Ray tracing configuration.

$sel:revisionId:AwsLambdaFunctionDetails', awsLambdaFunctionDetails_revisionId - The latest updated revision of the function or alias.

$sel:masterArn:AwsLambdaFunctionDetails', awsLambdaFunctionDetails_masterArn - For Lambda@Edge functions, the ARN of the master function.

awsLambdaFunctionDetails_memorySize :: Lens' AwsLambdaFunctionDetails (Maybe Int)

The memory that is allocated to the function.

awsLambdaFunctionDetails_runtime :: Lens' AwsLambdaFunctionDetails (Maybe Text)

The runtime environment for the Lambda function.

awsLambdaFunctionDetails_kmsKeyArn :: Lens' AwsLambdaFunctionDetails (Maybe Text)

The KMS key that is used to encrypt the function's environment variables. This key is only returned if you've configured a customer managed customer managed key.

awsLambdaFunctionDetails_environment :: Lens' AwsLambdaFunctionDetails (Maybe AwsLambdaFunctionEnvironment)

The function's environment variables.

awsLambdaFunctionDetails_deadLetterConfig :: Lens' AwsLambdaFunctionDetails (Maybe AwsLambdaFunctionDeadLetterConfig)

The function's dead letter queue.

awsLambdaFunctionDetails_role :: Lens' AwsLambdaFunctionDetails (Maybe Text)

The function's execution role.

awsLambdaFunctionDetails_vpcConfig :: Lens' AwsLambdaFunctionDetails (Maybe AwsLambdaFunctionVpcConfig)

The function's networking configuration.

awsLambdaFunctionDetails_version :: Lens' AwsLambdaFunctionDetails (Maybe Text)

The version of the Lambda function.

awsLambdaFunctionDetails_functionName :: Lens' AwsLambdaFunctionDetails (Maybe Text)

The name of the function.

awsLambdaFunctionDetails_code :: Lens' AwsLambdaFunctionDetails (Maybe AwsLambdaFunctionCode)

An AwsLambdaFunctionCode object.

awsLambdaFunctionDetails_layers :: Lens' AwsLambdaFunctionDetails (Maybe [AwsLambdaFunctionLayer])

The function's layers.

awsLambdaFunctionDetails_handler :: Lens' AwsLambdaFunctionDetails (Maybe Text)

The function that Lambda calls to begin executing your function.

awsLambdaFunctionDetails_timeout :: Lens' AwsLambdaFunctionDetails (Maybe Int)

The amount of time that Lambda allows a function to run before stopping it.

awsLambdaFunctionDetails_lastModified :: Lens' AwsLambdaFunctionDetails (Maybe Text)

Indicates when the function was last updated.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsLambdaFunctionDetails_codeSha256 :: Lens' AwsLambdaFunctionDetails (Maybe Text)

The SHA256 hash of the function's deployment package.

awsLambdaFunctionDetails_tracingConfig :: Lens' AwsLambdaFunctionDetails (Maybe AwsLambdaFunctionTracingConfig)

The function's X-Ray tracing configuration.

awsLambdaFunctionDetails_revisionId :: Lens' AwsLambdaFunctionDetails (Maybe Text)

The latest updated revision of the function or alias.

awsLambdaFunctionDetails_masterArn :: Lens' AwsLambdaFunctionDetails (Maybe Text)

For Lambda@Edge functions, the ARN of the master function.