Network.AWS.SecurityHub.Types.AwsRdsDbClusterSnapshotDetails

Description

Documentation

data AwsRdsDbClusterSnapshotDetails

Information about an Amazon RDS DB cluster snapshot.

See: newAwsRdsDbClusterSnapshotDetails smart constructor.

Constructors

AwsRdsDbClusterSnapshotDetails'

Fields

engineVersion :: Maybe Text

The version of the database engine to use.

status :: Maybe Text

The status of this DB cluster snapshot.

storageEncrypted :: Maybe Bool

Whether the DB cluster is encrypted.

dbClusterIdentifier :: Maybe Text

The DB cluster identifier.

masterUsername :: Maybe Text

The name of the master user for the DB cluster.

iamDatabaseAuthenticationEnabled :: Maybe Bool

Whether mapping of IAM accounts to database accounts is enabled.

vpcId :: Maybe Text

The VPC ID that is associated with the DB cluster snapshot.

dbClusterSnapshotIdentifier :: Maybe Text

The identifier of the DB cluster snapshot.

engine :: Maybe Text

The name of the database engine that you want to use for this DB instance.

licenseModel :: Maybe Text

The license model information for this DB cluster snapshot.

availabilityZones :: Maybe [Text]

A list of Availability Zones where instances in the DB cluster can be created.

snapshotType :: Maybe Text

The type of DB cluster snapshot.

kmsKeyId :: Maybe Text

The ARN of the KMS master key that is used to encrypt the database instances in the DB cluster.

snapshotCreateTime :: Maybe Text

Indicates when the snapshot was taken.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

allocatedStorage :: Maybe Int

Specifies the allocated storage size in gibibytes (GiB).

clusterCreateTime :: Maybe Text

Indicates when the DB cluster was created, in Universal Coordinated Time (UTC).

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

percentProgress :: Maybe Int

Specifies the percentage of the estimated data that has been transferred.

port :: Maybe Int

The port number on which the DB instances in the DB cluster accept connections.

newAwsRdsDbClusterSnapshotDetails :: AwsRdsDbClusterSnapshotDetails

Create a value of AwsRdsDbClusterSnapshotDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:AwsRdsDbClusterSnapshotDetails', awsRdsDbClusterSnapshotDetails_engineVersion - The version of the database engine to use.

$sel:status:AwsRdsDbClusterSnapshotDetails', awsRdsDbClusterSnapshotDetails_status - The status of this DB cluster snapshot.

$sel:storageEncrypted:AwsRdsDbClusterSnapshotDetails', awsRdsDbClusterSnapshotDetails_storageEncrypted - Whether the DB cluster is encrypted.

$sel:dbClusterIdentifier:AwsRdsDbClusterSnapshotDetails', awsRdsDbClusterSnapshotDetails_dbClusterIdentifier - The DB cluster identifier.

$sel:masterUsername:AwsRdsDbClusterSnapshotDetails', awsRdsDbClusterSnapshotDetails_masterUsername - The name of the master user for the DB cluster.

$sel:iamDatabaseAuthenticationEnabled:AwsRdsDbClusterSnapshotDetails', awsRdsDbClusterSnapshotDetails_iamDatabaseAuthenticationEnabled - Whether mapping of IAM accounts to database accounts is enabled.

$sel:vpcId:AwsRdsDbClusterSnapshotDetails', awsRdsDbClusterSnapshotDetails_vpcId - The VPC ID that is associated with the DB cluster snapshot.

$sel:dbClusterSnapshotIdentifier:AwsRdsDbClusterSnapshotDetails', awsRdsDbClusterSnapshotDetails_dbClusterSnapshotIdentifier - The identifier of the DB cluster snapshot.

$sel:engine:AwsRdsDbClusterSnapshotDetails', awsRdsDbClusterSnapshotDetails_engine - The name of the database engine that you want to use for this DB instance.

$sel:licenseModel:AwsRdsDbClusterSnapshotDetails', awsRdsDbClusterSnapshotDetails_licenseModel - The license model information for this DB cluster snapshot.

$sel:availabilityZones:AwsRdsDbClusterSnapshotDetails', awsRdsDbClusterSnapshotDetails_availabilityZones - A list of Availability Zones where instances in the DB cluster can be created.

$sel:snapshotType:AwsRdsDbClusterSnapshotDetails', awsRdsDbClusterSnapshotDetails_snapshotType - The type of DB cluster snapshot.

$sel:kmsKeyId:AwsRdsDbClusterSnapshotDetails', awsRdsDbClusterSnapshotDetails_kmsKeyId - The ARN of the KMS master key that is used to encrypt the database instances in the DB cluster.

$sel:snapshotCreateTime:AwsRdsDbClusterSnapshotDetails', awsRdsDbClusterSnapshotDetails_snapshotCreateTime - Indicates when the snapshot was taken.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:allocatedStorage:AwsRdsDbClusterSnapshotDetails', awsRdsDbClusterSnapshotDetails_allocatedStorage - Specifies the allocated storage size in gibibytes (GiB).

$sel:clusterCreateTime:AwsRdsDbClusterSnapshotDetails', awsRdsDbClusterSnapshotDetails_clusterCreateTime - Indicates when the DB cluster was created, in Universal Coordinated Time (UTC).

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:percentProgress:AwsRdsDbClusterSnapshotDetails', awsRdsDbClusterSnapshotDetails_percentProgress - Specifies the percentage of the estimated data that has been transferred.

$sel:port:AwsRdsDbClusterSnapshotDetails', awsRdsDbClusterSnapshotDetails_port - The port number on which the DB instances in the DB cluster accept connections.

awsRdsDbClusterSnapshotDetails_engineVersion :: Lens' AwsRdsDbClusterSnapshotDetails (Maybe Text)

The version of the database engine to use.

awsRdsDbClusterSnapshotDetails_status :: Lens' AwsRdsDbClusterSnapshotDetails (Maybe Text)

The status of this DB cluster snapshot.

awsRdsDbClusterSnapshotDetails_storageEncrypted :: Lens' AwsRdsDbClusterSnapshotDetails (Maybe Bool)

Whether the DB cluster is encrypted.

awsRdsDbClusterSnapshotDetails_dbClusterIdentifier :: Lens' AwsRdsDbClusterSnapshotDetails (Maybe Text)

The DB cluster identifier.

awsRdsDbClusterSnapshotDetails_masterUsername :: Lens' AwsRdsDbClusterSnapshotDetails (Maybe Text)

The name of the master user for the DB cluster.

awsRdsDbClusterSnapshotDetails_iamDatabaseAuthenticationEnabled :: Lens' AwsRdsDbClusterSnapshotDetails (Maybe Bool)

Whether mapping of IAM accounts to database accounts is enabled.

awsRdsDbClusterSnapshotDetails_vpcId :: Lens' AwsRdsDbClusterSnapshotDetails (Maybe Text)

The VPC ID that is associated with the DB cluster snapshot.

awsRdsDbClusterSnapshotDetails_dbClusterSnapshotIdentifier :: Lens' AwsRdsDbClusterSnapshotDetails (Maybe Text)

The identifier of the DB cluster snapshot.

awsRdsDbClusterSnapshotDetails_engine :: Lens' AwsRdsDbClusterSnapshotDetails (Maybe Text)

The name of the database engine that you want to use for this DB instance.

awsRdsDbClusterSnapshotDetails_licenseModel :: Lens' AwsRdsDbClusterSnapshotDetails (Maybe Text)

The license model information for this DB cluster snapshot.

awsRdsDbClusterSnapshotDetails_availabilityZones :: Lens' AwsRdsDbClusterSnapshotDetails (Maybe [Text])

A list of Availability Zones where instances in the DB cluster can be created.

awsRdsDbClusterSnapshotDetails_snapshotType :: Lens' AwsRdsDbClusterSnapshotDetails (Maybe Text)

The type of DB cluster snapshot.

awsRdsDbClusterSnapshotDetails_kmsKeyId :: Lens' AwsRdsDbClusterSnapshotDetails (Maybe Text)

The ARN of the KMS master key that is used to encrypt the database instances in the DB cluster.

awsRdsDbClusterSnapshotDetails_snapshotCreateTime :: Lens' AwsRdsDbClusterSnapshotDetails (Maybe Text)

Indicates when the snapshot was taken.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsRdsDbClusterSnapshotDetails_allocatedStorage :: Lens' AwsRdsDbClusterSnapshotDetails (Maybe Int)

Specifies the allocated storage size in gibibytes (GiB).

awsRdsDbClusterSnapshotDetails_clusterCreateTime :: Lens' AwsRdsDbClusterSnapshotDetails (Maybe Text)

Indicates when the DB cluster was created, in Universal Coordinated Time (UTC).

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsRdsDbClusterSnapshotDetails_percentProgress :: Lens' AwsRdsDbClusterSnapshotDetails (Maybe Int)

Specifies the percentage of the estimated data that has been transferred.

awsRdsDbClusterSnapshotDetails_port :: Lens' AwsRdsDbClusterSnapshotDetails (Maybe Int)

The port number on which the DB instances in the DB cluster accept connections.