Network.AWS.SecurityHub.Types.AwsRdsDbDomainMembership

Description

Documentation

data AwsRdsDbDomainMembership

Information about an Active Directory domain membership record associated with the DB instance.

See: newAwsRdsDbDomainMembership smart constructor.

Constructors

AwsRdsDbDomainMembership'

Fields

status :: Maybe Text

The status of the Active Directory Domain membership for the DB instance.

fqdn :: Maybe Text

The fully qualified domain name of the Active Directory domain.

domain :: Maybe Text

The identifier of the Active Directory domain.

iamRoleName :: Maybe Text

The name of the IAM role to use when making API calls to the Directory Service.

newAwsRdsDbDomainMembership :: AwsRdsDbDomainMembership

Create a value of AwsRdsDbDomainMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AwsRdsDbDomainMembership', awsRdsDbDomainMembership_status - The status of the Active Directory Domain membership for the DB instance.

$sel:fqdn:AwsRdsDbDomainMembership', awsRdsDbDomainMembership_fqdn - The fully qualified domain name of the Active Directory domain.

$sel:domain:AwsRdsDbDomainMembership', awsRdsDbDomainMembership_domain - The identifier of the Active Directory domain.

$sel:iamRoleName:AwsRdsDbDomainMembership', awsRdsDbDomainMembership_iamRoleName - The name of the IAM role to use when making API calls to the Directory Service.

awsRdsDbDomainMembership_status :: Lens' AwsRdsDbDomainMembership (Maybe Text)

The status of the Active Directory Domain membership for the DB instance.

awsRdsDbDomainMembership_fqdn :: Lens' AwsRdsDbDomainMembership (Maybe Text)

The fully qualified domain name of the Active Directory domain.

awsRdsDbDomainMembership_domain :: Lens' AwsRdsDbDomainMembership (Maybe Text)

The identifier of the Active Directory domain.

awsRdsDbDomainMembership_iamRoleName :: Lens' AwsRdsDbDomainMembership (Maybe Text)

The name of the IAM role to use when making API calls to the Directory Service.