Network.AWS.SecurityHub.Types.AwsRdsEventSubscriptionDetails

Description

Documentation

data AwsRdsEventSubscriptionDetails

Details about an Amazon RDS event notification subscription. The subscription allows Amazon RDS to post events to an SNS topic.

See: newAwsRdsEventSubscriptionDetails smart constructor.

Constructors

AwsRdsEventSubscriptionDetails'

Fields

status :: Maybe Text

The status of the event notification subscription.

Valid values: creating | modifying | deleting | active | no-permission | topic-not-exist

customerAwsId :: Maybe Text

The identifier of the event notification subscription.

custSubscriptionId :: Maybe Text

The identifier of the account that is associated with the event notification subscription.

snsTopicArn :: Maybe Text

The ARN of the SNS topic to post the event notifications to.

eventSubscriptionArn :: Maybe Text

The ARN of the event notification subscription.

enabled :: Maybe Bool

Whether the event notification subscription is enabled.

sourceType :: Maybe Text

The source type for the event notification subscription.

subscriptionCreationTime :: Maybe Text

The datetime when the event notification subscription was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

eventCategoriesList :: Maybe [Text]

The list of event categories for the event notification subscription.

sourceIdsList :: Maybe [Text]

A list of source identifiers for the event notification subscription.

newAwsRdsEventSubscriptionDetails :: AwsRdsEventSubscriptionDetails

Create a value of AwsRdsEventSubscriptionDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AwsRdsEventSubscriptionDetails', awsRdsEventSubscriptionDetails_status - The status of the event notification subscription.

Valid values: creating | modifying | deleting | active | no-permission | topic-not-exist

$sel:customerAwsId:AwsRdsEventSubscriptionDetails', awsRdsEventSubscriptionDetails_customerAwsId - The identifier of the event notification subscription.

$sel:custSubscriptionId:AwsRdsEventSubscriptionDetails', awsRdsEventSubscriptionDetails_custSubscriptionId - The identifier of the account that is associated with the event notification subscription.

$sel:snsTopicArn:AwsRdsEventSubscriptionDetails', awsRdsEventSubscriptionDetails_snsTopicArn - The ARN of the SNS topic to post the event notifications to.

$sel:eventSubscriptionArn:AwsRdsEventSubscriptionDetails', awsRdsEventSubscriptionDetails_eventSubscriptionArn - The ARN of the event notification subscription.

$sel:enabled:AwsRdsEventSubscriptionDetails', awsRdsEventSubscriptionDetails_enabled - Whether the event notification subscription is enabled.

$sel:sourceType:AwsRdsEventSubscriptionDetails', awsRdsEventSubscriptionDetails_sourceType - The source type for the event notification subscription.

$sel:subscriptionCreationTime:AwsRdsEventSubscriptionDetails', awsRdsEventSubscriptionDetails_subscriptionCreationTime - The datetime when the event notification subscription was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:eventCategoriesList:AwsRdsEventSubscriptionDetails', awsRdsEventSubscriptionDetails_eventCategoriesList - The list of event categories for the event notification subscription.

$sel:sourceIdsList:AwsRdsEventSubscriptionDetails', awsRdsEventSubscriptionDetails_sourceIdsList - A list of source identifiers for the event notification subscription.

awsRdsEventSubscriptionDetails_status :: Lens' AwsRdsEventSubscriptionDetails (Maybe Text)

The status of the event notification subscription.

Valid values: creating | modifying | deleting | active | no-permission | topic-not-exist

awsRdsEventSubscriptionDetails_customerAwsId :: Lens' AwsRdsEventSubscriptionDetails (Maybe Text)

The identifier of the event notification subscription.

awsRdsEventSubscriptionDetails_custSubscriptionId :: Lens' AwsRdsEventSubscriptionDetails (Maybe Text)

The identifier of the account that is associated with the event notification subscription.

awsRdsEventSubscriptionDetails_snsTopicArn :: Lens' AwsRdsEventSubscriptionDetails (Maybe Text)

The ARN of the SNS topic to post the event notifications to.

awsRdsEventSubscriptionDetails_eventSubscriptionArn :: Lens' AwsRdsEventSubscriptionDetails (Maybe Text)

The ARN of the event notification subscription.

awsRdsEventSubscriptionDetails_enabled :: Lens' AwsRdsEventSubscriptionDetails (Maybe Bool)

Whether the event notification subscription is enabled.

awsRdsEventSubscriptionDetails_sourceType :: Lens' AwsRdsEventSubscriptionDetails (Maybe Text)

The source type for the event notification subscription.

awsRdsEventSubscriptionDetails_subscriptionCreationTime :: Lens' AwsRdsEventSubscriptionDetails (Maybe Text)

The datetime when the event notification subscription was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsRdsEventSubscriptionDetails_eventCategoriesList :: Lens' AwsRdsEventSubscriptionDetails (Maybe [Text])

The list of event categories for the event notification subscription.

awsRdsEventSubscriptionDetails_sourceIdsList :: Lens' AwsRdsEventSubscriptionDetails (Maybe [Text])

A list of source identifiers for the event notification subscription.