Network.AWS.SecurityHub.Types.AwsRedshiftClusterRestoreStatus

Description

Documentation

data AwsRedshiftClusterRestoreStatus

Information about the status of a cluster restore action. It only applies if the cluster was created by restoring a snapshot.

See: newAwsRedshiftClusterRestoreStatus smart constructor.

Constructors

AwsRedshiftClusterRestoreStatus'

Fields

status :: Maybe Text

The status of the restore action.

Valid values: starting | restoring | completed | failed

estimatedTimeToCompletionInSeconds :: Maybe Integer

The estimate of the time remaining before the restore is complete. Returns 0 for a completed restore.

This field is only updated when you restore to DC2 and DS2 node types.

currentRestoreRateInMegaBytesPerSecond :: Maybe Double

The number of megabytes per second being transferred from the backup storage. Returns the average rate for a completed backup.

This field is only updated when you restore to DC2 and DS2 node types.

progressInMegaBytes :: Maybe Integer

The number of megabytes that were transferred from snapshot storage.

This field is only updated when you restore to DC2 and DS2 node types.

elapsedTimeInSeconds :: Maybe Integer

The amount of time an in-progress restore has been running, or the amount of time it took a completed restore to finish.

This field is only updated when you restore to DC2 and DS2 node types.

snapshotSizeInMegaBytes :: Maybe Integer

The size of the set of snapshot data that was used to restore the cluster.

This field is only updated when you restore to DC2 and DS2 node types.

newAwsRedshiftClusterRestoreStatus :: AwsRedshiftClusterRestoreStatus

Create a value of AwsRedshiftClusterRestoreStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AwsRedshiftClusterRestoreStatus', awsRedshiftClusterRestoreStatus_status - The status of the restore action.

Valid values: starting | restoring | completed | failed

$sel:estimatedTimeToCompletionInSeconds:AwsRedshiftClusterRestoreStatus', awsRedshiftClusterRestoreStatus_estimatedTimeToCompletionInSeconds - The estimate of the time remaining before the restore is complete. Returns 0 for a completed restore.

This field is only updated when you restore to DC2 and DS2 node types.

$sel:currentRestoreRateInMegaBytesPerSecond:AwsRedshiftClusterRestoreStatus', awsRedshiftClusterRestoreStatus_currentRestoreRateInMegaBytesPerSecond - The number of megabytes per second being transferred from the backup storage. Returns the average rate for a completed backup.

This field is only updated when you restore to DC2 and DS2 node types.

$sel:progressInMegaBytes:AwsRedshiftClusterRestoreStatus', awsRedshiftClusterRestoreStatus_progressInMegaBytes - The number of megabytes that were transferred from snapshot storage.

This field is only updated when you restore to DC2 and DS2 node types.

$sel:elapsedTimeInSeconds:AwsRedshiftClusterRestoreStatus', awsRedshiftClusterRestoreStatus_elapsedTimeInSeconds - The amount of time an in-progress restore has been running, or the amount of time it took a completed restore to finish.

This field is only updated when you restore to DC2 and DS2 node types.

$sel:snapshotSizeInMegaBytes:AwsRedshiftClusterRestoreStatus', awsRedshiftClusterRestoreStatus_snapshotSizeInMegaBytes - The size of the set of snapshot data that was used to restore the cluster.

This field is only updated when you restore to DC2 and DS2 node types.

awsRedshiftClusterRestoreStatus_status :: Lens' AwsRedshiftClusterRestoreStatus (Maybe Text)

The status of the restore action.

Valid values: starting | restoring | completed | failed

awsRedshiftClusterRestoreStatus_estimatedTimeToCompletionInSeconds :: Lens' AwsRedshiftClusterRestoreStatus (Maybe Integer)

The estimate of the time remaining before the restore is complete. Returns 0 for a completed restore.

This field is only updated when you restore to DC2 and DS2 node types.

awsRedshiftClusterRestoreStatus_currentRestoreRateInMegaBytesPerSecond :: Lens' AwsRedshiftClusterRestoreStatus (Maybe Double)

The number of megabytes per second being transferred from the backup storage. Returns the average rate for a completed backup.

This field is only updated when you restore to DC2 and DS2 node types.

awsRedshiftClusterRestoreStatus_progressInMegaBytes :: Lens' AwsRedshiftClusterRestoreStatus (Maybe Integer)

The number of megabytes that were transferred from snapshot storage.

This field is only updated when you restore to DC2 and DS2 node types.

awsRedshiftClusterRestoreStatus_elapsedTimeInSeconds :: Lens' AwsRedshiftClusterRestoreStatus (Maybe Integer)

The amount of time an in-progress restore has been running, or the amount of time it took a completed restore to finish.

This field is only updated when you restore to DC2 and DS2 node types.

awsRedshiftClusterRestoreStatus_snapshotSizeInMegaBytes :: Lens' AwsRedshiftClusterRestoreStatus (Maybe Integer)

The size of the set of snapshot data that was used to restore the cluster.

This field is only updated when you restore to DC2 and DS2 node types.