Network.AWS.SecurityHub.Types.AwsS3AccountPublicAccessBlockDetails

Description

Documentation

data AwsS3AccountPublicAccessBlockDetails

provides information about the Amazon S3 Public Access Block configuration for accounts.

See: newAwsS3AccountPublicAccessBlockDetails smart constructor.

Constructors

AwsS3AccountPublicAccessBlockDetails'

Fields

ignorePublicAcls :: Maybe Bool

Indicates whether Amazon S3 ignores public ACLs that are associated with an S3 bucket.

blockPublicAcls :: Maybe Bool

Indicates whether to reject calls to update an S3 bucket if the calls include a public access control list (ACL).

restrictPublicBuckets :: Maybe Bool

Indicates whether to restrict access to an access point or S3 bucket that has a public policy to only Amazon Web Services service principals and authorized users within the S3 bucket owner's account.

blockPublicPolicy :: Maybe Bool

Indicates whether to reject calls to update the access policy for an S3 bucket or access point if the policy allows public access.

newAwsS3AccountPublicAccessBlockDetails :: AwsS3AccountPublicAccessBlockDetails

Create a value of AwsS3AccountPublicAccessBlockDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ignorePublicAcls:AwsS3AccountPublicAccessBlockDetails', awsS3AccountPublicAccessBlockDetails_ignorePublicAcls - Indicates whether Amazon S3 ignores public ACLs that are associated with an S3 bucket.

$sel:blockPublicAcls:AwsS3AccountPublicAccessBlockDetails', awsS3AccountPublicAccessBlockDetails_blockPublicAcls - Indicates whether to reject calls to update an S3 bucket if the calls include a public access control list (ACL).

$sel:restrictPublicBuckets:AwsS3AccountPublicAccessBlockDetails', awsS3AccountPublicAccessBlockDetails_restrictPublicBuckets - Indicates whether to restrict access to an access point or S3 bucket that has a public policy to only Amazon Web Services service principals and authorized users within the S3 bucket owner's account.

$sel:blockPublicPolicy:AwsS3AccountPublicAccessBlockDetails', awsS3AccountPublicAccessBlockDetails_blockPublicPolicy - Indicates whether to reject calls to update the access policy for an S3 bucket or access point if the policy allows public access.

awsS3AccountPublicAccessBlockDetails_ignorePublicAcls :: Lens' AwsS3AccountPublicAccessBlockDetails (Maybe Bool)

Indicates whether Amazon S3 ignores public ACLs that are associated with an S3 bucket.

awsS3AccountPublicAccessBlockDetails_blockPublicAcls :: Lens' AwsS3AccountPublicAccessBlockDetails (Maybe Bool)

Indicates whether to reject calls to update an S3 bucket if the calls include a public access control list (ACL).

awsS3AccountPublicAccessBlockDetails_restrictPublicBuckets :: Lens' AwsS3AccountPublicAccessBlockDetails (Maybe Bool)

Indicates whether to restrict access to an access point or S3 bucket that has a public policy to only Amazon Web Services service principals and authorized users within the S3 bucket owner's account.

awsS3AccountPublicAccessBlockDetails_blockPublicPolicy :: Lens' AwsS3AccountPublicAccessBlockDetails (Maybe Bool)

Indicates whether to reject calls to update the access policy for an S3 bucket or access point if the policy allows public access.