Network.AWS.SecurityHub.Types.AwsS3BucketDetails

Description

Documentation

data AwsS3BucketDetails

The details of an Amazon S3 bucket.

See: newAwsS3BucketDetails smart constructor.

Constructors

AwsS3BucketDetails'

Fields

createdAt :: Maybe Text

Indicates when the S3 bucket was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

ownerAccountId :: Maybe Text

The Amazon Web Services account identifier of the account that owns the S3 bucket.

bucketLoggingConfiguration :: Maybe AwsS3BucketLoggingConfiguration

The logging configuration for the S3 bucket.

ownerName :: Maybe Text

The display name of the owner of the S3 bucket.

accessControlList :: Maybe Text

The access control list for the S3 bucket.

ownerId :: Maybe Text

The canonical user ID of the owner of the S3 bucket.

publicAccessBlockConfiguration :: Maybe AwsS3AccountPublicAccessBlockDetails

Provides information about the Amazon S3 Public Access Block configuration for the S3 bucket.

bucketWebsiteConfiguration :: Maybe AwsS3BucketWebsiteConfiguration

The website configuration parameters for the S3 bucket.

bucketLifecycleConfiguration :: Maybe AwsS3BucketBucketLifecycleConfigurationDetails

The lifecycle configuration for objects in the S3 bucket.

bucketNotificationConfiguration :: Maybe AwsS3BucketNotificationConfiguration

The notification configuration for the S3 bucket.

serverSideEncryptionConfiguration :: Maybe AwsS3BucketServerSideEncryptionConfiguration

The encryption rules that are applied to the S3 bucket.

newAwsS3BucketDetails :: AwsS3BucketDetails

Create a value of AwsS3BucketDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAt:AwsS3BucketDetails', awsS3BucketDetails_createdAt - Indicates when the S3 bucket was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:ownerAccountId:AwsS3BucketDetails', awsS3BucketDetails_ownerAccountId - The Amazon Web Services account identifier of the account that owns the S3 bucket.

$sel:bucketLoggingConfiguration:AwsS3BucketDetails', awsS3BucketDetails_bucketLoggingConfiguration - The logging configuration for the S3 bucket.

$sel:ownerName:AwsS3BucketDetails', awsS3BucketDetails_ownerName - The display name of the owner of the S3 bucket.

$sel:accessControlList:AwsS3BucketDetails', awsS3BucketDetails_accessControlList - The access control list for the S3 bucket.

$sel:ownerId:AwsS3BucketDetails', awsS3BucketDetails_ownerId - The canonical user ID of the owner of the S3 bucket.

$sel:publicAccessBlockConfiguration:AwsS3BucketDetails', awsS3BucketDetails_publicAccessBlockConfiguration - Provides information about the Amazon S3 Public Access Block configuration for the S3 bucket.

$sel:bucketWebsiteConfiguration:AwsS3BucketDetails', awsS3BucketDetails_bucketWebsiteConfiguration - The website configuration parameters for the S3 bucket.

$sel:bucketLifecycleConfiguration:AwsS3BucketDetails', awsS3BucketDetails_bucketLifecycleConfiguration - The lifecycle configuration for objects in the S3 bucket.

$sel:bucketNotificationConfiguration:AwsS3BucketDetails', awsS3BucketDetails_bucketNotificationConfiguration - The notification configuration for the S3 bucket.

$sel:serverSideEncryptionConfiguration:AwsS3BucketDetails', awsS3BucketDetails_serverSideEncryptionConfiguration - The encryption rules that are applied to the S3 bucket.

awsS3BucketDetails_createdAt :: Lens' AwsS3BucketDetails (Maybe Text)

Indicates when the S3 bucket was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsS3BucketDetails_ownerAccountId :: Lens' AwsS3BucketDetails (Maybe Text)

The Amazon Web Services account identifier of the account that owns the S3 bucket.

awsS3BucketDetails_bucketLoggingConfiguration :: Lens' AwsS3BucketDetails (Maybe AwsS3BucketLoggingConfiguration)

The logging configuration for the S3 bucket.

awsS3BucketDetails_ownerName :: Lens' AwsS3BucketDetails (Maybe Text)

The display name of the owner of the S3 bucket.

awsS3BucketDetails_accessControlList :: Lens' AwsS3BucketDetails (Maybe Text)

The access control list for the S3 bucket.

awsS3BucketDetails_ownerId :: Lens' AwsS3BucketDetails (Maybe Text)

The canonical user ID of the owner of the S3 bucket.

awsS3BucketDetails_publicAccessBlockConfiguration :: Lens' AwsS3BucketDetails (Maybe AwsS3AccountPublicAccessBlockDetails)

Provides information about the Amazon S3 Public Access Block configuration for the S3 bucket.

awsS3BucketDetails_bucketWebsiteConfiguration :: Lens' AwsS3BucketDetails (Maybe AwsS3BucketWebsiteConfiguration)

The website configuration parameters for the S3 bucket.

awsS3BucketDetails_bucketLifecycleConfiguration :: Lens' AwsS3BucketDetails (Maybe AwsS3BucketBucketLifecycleConfigurationDetails)

The lifecycle configuration for objects in the S3 bucket.

awsS3BucketDetails_bucketNotificationConfiguration :: Lens' AwsS3BucketDetails (Maybe AwsS3BucketNotificationConfiguration)

The notification configuration for the S3 bucket.

awsS3BucketDetails_serverSideEncryptionConfiguration :: Lens' AwsS3BucketDetails (Maybe AwsS3BucketServerSideEncryptionConfiguration)

The encryption rules that are applied to the S3 bucket.