Network.AWS.SecurityHub.Types.AwsS3BucketWebsiteConfigurationRoutingRuleCondition

Description

Documentation

data AwsS3BucketWebsiteConfigurationRoutingRuleCondition

The condition that must be met in order to apply the routing rule.

See: newAwsS3BucketWebsiteConfigurationRoutingRuleCondition smart constructor.

Constructors

AwsS3BucketWebsiteConfigurationRoutingRuleCondition'

Fields

keyPrefixEquals :: Maybe Text

Indicates to redirect the request if the key prefix matches this value.

httpErrorCodeReturnedEquals :: Maybe Text

Indicates to redirect the request if the HTTP error code matches this value.

newAwsS3BucketWebsiteConfigurationRoutingRuleCondition :: AwsS3BucketWebsiteConfigurationRoutingRuleCondition

Create a value of AwsS3BucketWebsiteConfigurationRoutingRuleCondition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyPrefixEquals:AwsS3BucketWebsiteConfigurationRoutingRuleCondition', awsS3BucketWebsiteConfigurationRoutingRuleCondition_keyPrefixEquals - Indicates to redirect the request if the key prefix matches this value.

$sel:httpErrorCodeReturnedEquals:AwsS3BucketWebsiteConfigurationRoutingRuleCondition', awsS3BucketWebsiteConfigurationRoutingRuleCondition_httpErrorCodeReturnedEquals - Indicates to redirect the request if the HTTP error code matches this value.

awsS3BucketWebsiteConfigurationRoutingRuleCondition_keyPrefixEquals :: Lens' AwsS3BucketWebsiteConfigurationRoutingRuleCondition (Maybe Text)

Indicates to redirect the request if the key prefix matches this value.

awsS3BucketWebsiteConfigurationRoutingRuleCondition_httpErrorCodeReturnedEquals :: Lens' AwsS3BucketWebsiteConfigurationRoutingRuleCondition (Maybe Text)

Indicates to redirect the request if the HTTP error code matches this value.