Network.AWS.SecurityHub.Types.AwsWafRegionalRateBasedRuleDetails

Description

Documentation

data AwsWafRegionalRateBasedRuleDetails

contains details about a rate-based rule for Regional resources. A rate-based rule provides settings to indicate when to allow, block, or count a request. Rate-based rules include the number of requests that arrive over a specified period of time.

See: newAwsWafRegionalRateBasedRuleDetails smart constructor.

Constructors

AwsWafRegionalRateBasedRuleDetails'

Fields

rateLimit :: Maybe Integer

The maximum number of requests that have an identical value for the field specified in RateKey that are allowed within a five-minute period. If the number of requests exceeds RateLimit and the other predicates specified in the rule are met, WAF triggers the action for the rule.

rateKey :: Maybe Text

The field that WAF uses to determine whether requests are likely arriving from single source and are subject to rate monitoring.

metricName :: Maybe Text

The name of the metrics for the rate-based rule.

ruleId :: Maybe Text

The unique identifier for the rate-based rule.

name :: Maybe Text

The name of the rate-based rule.

matchPredicates :: Maybe [AwsWafRegionalRateBasedRuleMatchPredicate]

The predicates to include in the rate-based rule.

newAwsWafRegionalRateBasedRuleDetails :: AwsWafRegionalRateBasedRuleDetails

Create a value of AwsWafRegionalRateBasedRuleDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rateLimit:AwsWafRegionalRateBasedRuleDetails', awsWafRegionalRateBasedRuleDetails_rateLimit - The maximum number of requests that have an identical value for the field specified in RateKey that are allowed within a five-minute period. If the number of requests exceeds RateLimit and the other predicates specified in the rule are met, WAF triggers the action for the rule.

$sel:rateKey:AwsWafRegionalRateBasedRuleDetails', awsWafRegionalRateBasedRuleDetails_rateKey - The field that WAF uses to determine whether requests are likely arriving from single source and are subject to rate monitoring.

$sel:metricName:AwsWafRegionalRateBasedRuleDetails', awsWafRegionalRateBasedRuleDetails_metricName - The name of the metrics for the rate-based rule.

$sel:ruleId:AwsWafRegionalRateBasedRuleDetails', awsWafRegionalRateBasedRuleDetails_ruleId - The unique identifier for the rate-based rule.

$sel:name:AwsWafRegionalRateBasedRuleDetails', awsWafRegionalRateBasedRuleDetails_name - The name of the rate-based rule.

$sel:matchPredicates:AwsWafRegionalRateBasedRuleDetails', awsWafRegionalRateBasedRuleDetails_matchPredicates - The predicates to include in the rate-based rule.

awsWafRegionalRateBasedRuleDetails_rateLimit :: Lens' AwsWafRegionalRateBasedRuleDetails (Maybe Integer)

The maximum number of requests that have an identical value for the field specified in RateKey that are allowed within a five-minute period. If the number of requests exceeds RateLimit and the other predicates specified in the rule are met, WAF triggers the action for the rule.

awsWafRegionalRateBasedRuleDetails_rateKey :: Lens' AwsWafRegionalRateBasedRuleDetails (Maybe Text)

The field that WAF uses to determine whether requests are likely arriving from single source and are subject to rate monitoring.

awsWafRegionalRateBasedRuleDetails_metricName :: Lens' AwsWafRegionalRateBasedRuleDetails (Maybe Text)

The name of the metrics for the rate-based rule.

awsWafRegionalRateBasedRuleDetails_ruleId :: Lens' AwsWafRegionalRateBasedRuleDetails (Maybe Text)

The unique identifier for the rate-based rule.

awsWafRegionalRateBasedRuleDetails_name :: Lens' AwsWafRegionalRateBasedRuleDetails (Maybe Text)

The name of the rate-based rule.

awsWafRegionalRateBasedRuleDetails_matchPredicates :: Lens' AwsWafRegionalRateBasedRuleDetails (Maybe [AwsWafRegionalRateBasedRuleMatchPredicate])

The predicates to include in the rate-based rule.