Network.AWS.SecurityHub.Types.AwsWafWebAclDetails

Description

Documentation

data AwsWafWebAclDetails

Details about an WAF WebACL.

See: newAwsWafWebAclDetails smart constructor.

Constructors

AwsWafWebAclDetails'

Fields

rules :: Maybe [AwsWafWebAclRule]

An array that contains the action for each rule in a WebACL, the priority of the rule, and the ID of the rule.

webAclId :: Maybe Text

A unique identifier for a WebACL.

name :: Maybe Text

A friendly name or description of the WebACL. You can't change the name of a WebACL after you create it.

defaultAction :: Maybe Text

The action to perform if none of the rules contained in the WebACL match.

newAwsWafWebAclDetails :: AwsWafWebAclDetails

Create a value of AwsWafWebAclDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rules:AwsWafWebAclDetails', awsWafWebAclDetails_rules - An array that contains the action for each rule in a WebACL, the priority of the rule, and the ID of the rule.

$sel:webAclId:AwsWafWebAclDetails', awsWafWebAclDetails_webAclId - A unique identifier for a WebACL.

$sel:name:AwsWafWebAclDetails', awsWafWebAclDetails_name - A friendly name or description of the WebACL. You can't change the name of a WebACL after you create it.

$sel:defaultAction:AwsWafWebAclDetails', awsWafWebAclDetails_defaultAction - The action to perform if none of the rules contained in the WebACL match.

awsWafWebAclDetails_rules :: Lens' AwsWafWebAclDetails (Maybe [AwsWafWebAclRule])

An array that contains the action for each rule in a WebACL, the priority of the rule, and the ID of the rule.

awsWafWebAclDetails_webAclId :: Lens' AwsWafWebAclDetails (Maybe Text)

A unique identifier for a WebACL.

awsWafWebAclDetails_name :: Lens' AwsWafWebAclDetails (Maybe Text)

A friendly name or description of the WebACL. You can't change the name of a WebACL after you create it.

awsWafWebAclDetails_defaultAction :: Lens' AwsWafWebAclDetails (Maybe Text)

The action to perform if none of the rules contained in the WebACL match.