Network.AWS.SecurityHub.Types.ClassificationResult

Description

Documentation

data ClassificationResult

Details about the sensitive data that was detected on the resource.

See: newClassificationResult smart constructor.

Constructors

ClassificationResult'

Fields

status :: Maybe ClassificationStatus

The current status of the sensitive data detection.

sensitiveData :: Maybe [SensitiveDataResult]

Provides details about sensitive data that was identified based on built-in configuration.

mimeType :: Maybe Text

The type of content that the finding applies to.

sizeClassified :: Maybe Integer

The total size in bytes of the affected data.

additionalOccurrences :: Maybe Bool

Indicates whether there are additional occurrences of sensitive data that are not included in the finding. This occurs when the number of occurrences exceeds the maximum that can be included.

customDataIdentifiers :: Maybe CustomDataIdentifiersResult

Provides details about sensitive data that was identified based on customer-defined configuration.

newClassificationResult :: ClassificationResult

Create a value of ClassificationResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ClassificationResult', classificationResult_status - The current status of the sensitive data detection.

$sel:sensitiveData:ClassificationResult', classificationResult_sensitiveData - Provides details about sensitive data that was identified based on built-in configuration.

$sel:mimeType:ClassificationResult', classificationResult_mimeType - The type of content that the finding applies to.

$sel:sizeClassified:ClassificationResult', classificationResult_sizeClassified - The total size in bytes of the affected data.

$sel:additionalOccurrences:ClassificationResult', classificationResult_additionalOccurrences - Indicates whether there are additional occurrences of sensitive data that are not included in the finding. This occurs when the number of occurrences exceeds the maximum that can be included.

$sel:customDataIdentifiers:ClassificationResult', classificationResult_customDataIdentifiers - Provides details about sensitive data that was identified based on customer-defined configuration.

classificationResult_status :: Lens' ClassificationResult (Maybe ClassificationStatus)

The current status of the sensitive data detection.

classificationResult_sensitiveData :: Lens' ClassificationResult (Maybe [SensitiveDataResult])

Provides details about sensitive data that was identified based on built-in configuration.

classificationResult_mimeType :: Lens' ClassificationResult (Maybe Text)

The type of content that the finding applies to.

classificationResult_sizeClassified :: Lens' ClassificationResult (Maybe Integer)

The total size in bytes of the affected data.

classificationResult_additionalOccurrences :: Lens' ClassificationResult (Maybe Bool)

Indicates whether there are additional occurrences of sensitive data that are not included in the finding. This occurs when the number of occurrences exceeds the maximum that can be included.

classificationResult_customDataIdentifiers :: Lens' ClassificationResult (Maybe CustomDataIdentifiersResult)

Provides details about sensitive data that was identified based on customer-defined configuration.