Network.AWS.SecurityHub.Types.Insight

Description

Documentation

data Insight

Contains information about a Security Hub insight.

See: newInsight smart constructor.

Constructors

Insight'

Fields

insightArn :: Text

The ARN of a Security Hub insight.

name :: Text

The name of a Security Hub insight.

filters :: AwsSecurityFindingFilters

One or more attributes used to filter the findings included in the insight. The insight only includes findings that match the criteria defined in the filters.

groupByAttribute :: Text

The grouping attribute for the insight's findings. Indicates how to group the matching findings, and identifies the type of item that the insight applies to. For example, if an insight is grouped by resource identifier, then the insight produces a list of resource identifiers.

newInsight

Arguments

:: Text
$sel:insightArn:Insight'

-> Text
$sel:name:Insight'

-> AwsSecurityFindingFilters
$sel:filters:Insight'

-> Text
$sel:groupByAttribute:Insight'

-> Insight

Create a value of Insight with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:insightArn:Insight', insight_insightArn - The ARN of a Security Hub insight.

$sel:name:Insight', insight_name - The name of a Security Hub insight.

$sel:filters:Insight', insight_filters - One or more attributes used to filter the findings included in the insight. The insight only includes findings that match the criteria defined in the filters.

$sel:groupByAttribute:Insight', insight_groupByAttribute - The grouping attribute for the insight's findings. Indicates how to group the matching findings, and identifies the type of item that the insight applies to. For example, if an insight is grouped by resource identifier, then the insight produces a list of resource identifiers.

insight_insightArn :: Lens' Insight Text

The ARN of a Security Hub insight.

insight_name :: Lens' Insight Text

The name of a Security Hub insight.

insight_filters :: Lens' Insight AwsSecurityFindingFilters

One or more attributes used to filter the findings included in the insight. The insight only includes findings that match the criteria defined in the filters.

insight_groupByAttribute :: Lens' Insight Text

The grouping attribute for the insight's findings. Indicates how to group the matching findings, and identifies the type of item that the insight applies to. For example, if an insight is grouped by resource identifier, then the insight produces a list of resource identifiers.