Network.AWS.SecurityHub.Types.Product

Description

Documentation

data Product

Contains details about a product.

See: newProduct smart constructor.

Constructors

Product'

Fields

productName :: Maybe Text

The name of the product.

productSubscriptionResourcePolicy :: Maybe Text

The resource policy associated with the product.

companyName :: Maybe Text

The name of the company that provides the product.

categories :: Maybe [Text]

The categories assigned to the product.

marketplaceUrl :: Maybe Text

For integrations with Amazon Web Services services, the Amazon Web Services Console URL from which to activate the service.

For integrations with third-party products, the Amazon Web Services Marketplace URL from which to subscribe to or purchase the product.

activationUrl :: Maybe Text

The URL to the service or product documentation about the integration with Security Hub, including how to activate the integration.

integrationTypes :: Maybe [IntegrationType]

The types of integration that the product supports. Available values are the following.

SEND_FINDINGS_TO_SECURITY_HUB - The integration sends findings to Security Hub.

RECEIVE_FINDINGS_FROM_SECURITY_HUB - The integration receives findings from Security Hub.

UPDATE_FINDINGS_IN_SECURITY_HUB - The integration does not send new findings to Security Hub, but does make updates to the findings that it receives from Security Hub.

description :: Maybe Text

A description of the product.

productArn :: Text

The ARN assigned to the product.

newProduct

Arguments

:: Text
$sel:productArn:Product'

-> Product

Create a value of Product with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:productName:Product', product_productName - The name of the product.

$sel:productSubscriptionResourcePolicy:Product', product_productSubscriptionResourcePolicy - The resource policy associated with the product.

$sel:companyName:Product', product_companyName - The name of the company that provides the product.

$sel:categories:Product', product_categories - The categories assigned to the product.

$sel:marketplaceUrl:Product', product_marketplaceUrl - For integrations with Amazon Web Services services, the Amazon Web Services Console URL from which to activate the service.

For integrations with third-party products, the Amazon Web Services Marketplace URL from which to subscribe to or purchase the product.

$sel:activationUrl:Product', product_activationUrl - The URL to the service or product documentation about the integration with Security Hub, including how to activate the integration.

$sel:integrationTypes:Product', product_integrationTypes - The types of integration that the product supports. Available values are the following.

SEND_FINDINGS_TO_SECURITY_HUB - The integration sends findings to Security Hub.

RECEIVE_FINDINGS_FROM_SECURITY_HUB - The integration receives findings from Security Hub.

UPDATE_FINDINGS_IN_SECURITY_HUB - The integration does not send new findings to Security Hub, but does make updates to the findings that it receives from Security Hub.

$sel:description:Product', product_description - A description of the product.

$sel:productArn:Product', product_productArn - The ARN assigned to the product.

product_productName :: Lens' Product (Maybe Text)

The name of the product.

product_productSubscriptionResourcePolicy :: Lens' Product (Maybe Text)

The resource policy associated with the product.

product_companyName :: Lens' Product (Maybe Text)

The name of the company that provides the product.

product_categories :: Lens' Product (Maybe [Text])

The categories assigned to the product.

product_marketplaceUrl :: Lens' Product (Maybe Text)

For integrations with Amazon Web Services services, the Amazon Web Services Console URL from which to activate the service.

For integrations with third-party products, the Amazon Web Services Marketplace URL from which to subscribe to or purchase the product.

product_activationUrl :: Lens' Product (Maybe Text)

The URL to the service or product documentation about the integration with Security Hub, including how to activate the integration.

product_integrationTypes :: Lens' Product (Maybe [IntegrationType])

The types of integration that the product supports. Available values are the following.

SEND_FINDINGS_TO_SECURITY_HUB - The integration sends findings to Security Hub.

RECEIVE_FINDINGS_FROM_SECURITY_HUB - The integration receives findings from Security Hub.

UPDATE_FINDINGS_IN_SECURITY_HUB - The integration does not send new findings to Security Hub, but does make updates to the findings that it receives from Security Hub.

product_description :: Lens' Product (Maybe Text)

A description of the product.

product_productArn :: Lens' Product Text

The ARN assigned to the product.