Network.AWS.SecurityHub.Types.Record

Description

Documentation

data Record

An occurrence of sensitive data in an Apache Avro object container or an Apache Parquet file.

See: newRecord smart constructor.

Constructors

Record'

Fields

jsonPath :: Maybe Text

The path, as a JSONPath expression, to the field in the record that contains the data. If the field name is longer than 20 characters, it is truncated. If the path is longer than 250 characters, it is truncated.

recordIndex :: Maybe Integer

The record index, starting from 0, for the record that contains the data.

newRecord :: Record

Create a value of Record with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jsonPath:Record', record_jsonPath - The path, as a JSONPath expression, to the field in the record that contains the data. If the field name is longer than 20 characters, it is truncated. If the path is longer than 250 characters, it is truncated.

$sel:recordIndex:Record', record_recordIndex - The record index, starting from 0, for the record that contains the data.

record_jsonPath :: Lens' Record (Maybe Text)

The path, as a JSONPath expression, to the field in the record that contains the data. If the field name is longer than 20 characters, it is truncated. If the path is longer than 250 characters, it is truncated.

record_recordIndex :: Lens' Record (Maybe Integer)

The record index, starting from 0, for the record that contains the data.