Network.AWS.SecurityHub.Types.Severity

Description

Documentation

data Severity

The severity of the finding.

The finding provider can provide the initial severity. The finding provider can only update the severity if it has not been updated using BatchUpdateFindings.

The finding must have either Label or Normalized populated. If only one of these attributes is populated, then Security Hub automatically populates the other one. If neither attribute is populated, then the finding is invalid. Label is the preferred attribute.

See: newSeverity smart constructor.

Constructors

Severity'

Fields

product :: Maybe Double

Deprecated. This attribute is being deprecated. Instead of providing Product, provide Original.

The native severity as defined by the Amazon Web Services service or integrated partner product that generated the finding.

label :: Maybe SeverityLabel

The severity value of the finding. The allowed values are the following.

INFORMATIONAL - No issue was found.

LOW - The issue does not require action on its own.

MEDIUM - The issue must be addressed but not urgently.

HIGH - The issue must be addressed as a priority.

CRITICAL - The issue must be remediated immediately to avoid it escalating.

If you provide Normalized and do not provide Label, then Label is set automatically as follows.

0 - INFORMATIONAL

1–39 - LOW

40–69 - MEDIUM

70–89 - HIGH

90–100 - CRITICAL

original :: Maybe Text

The native severity from the finding product that generated the finding.

normalized :: Maybe Int

Deprecated. The normalized severity of a finding. This attribute is being deprecated. Instead of providing Normalized, provide Label.

If you provide Label and do not provide Normalized, then Normalized is set automatically as follows.

INFORMATIONAL - 0

LOW - 1

MEDIUM - 40

HIGH - 70

CRITICAL - 90

newSeverity :: Severity

Create a value of Severity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:product:Severity', severity_product - Deprecated. This attribute is being deprecated. Instead of providing Product, provide Original.

The native severity as defined by the Amazon Web Services service or integrated partner product that generated the finding.

$sel:label:Severity', severity_label - The severity value of the finding. The allowed values are the following.

INFORMATIONAL - No issue was found.

LOW - The issue does not require action on its own.

MEDIUM - The issue must be addressed but not urgently.

HIGH - The issue must be addressed as a priority.

CRITICAL - The issue must be remediated immediately to avoid it escalating.

If you provide Normalized and do not provide Label, then Label is set automatically as follows.

0 - INFORMATIONAL

1–39 - LOW

40–69 - MEDIUM

70–89 - HIGH

90–100 - CRITICAL

$sel:original:Severity', severity_original - The native severity from the finding product that generated the finding.

$sel:normalized:Severity', severity_normalized - Deprecated. The normalized severity of a finding. This attribute is being deprecated. Instead of providing Normalized, provide Label.

If you provide Label and do not provide Normalized, then Normalized is set automatically as follows.

INFORMATIONAL - 0

LOW - 1

MEDIUM - 40

HIGH - 70

CRITICAL - 90

severity_product :: Lens' Severity (Maybe Double)

Deprecated. This attribute is being deprecated. Instead of providing Product, provide Original.

The native severity as defined by the Amazon Web Services service or integrated partner product that generated the finding.

severity_label :: Lens' Severity (Maybe SeverityLabel)

The severity value of the finding. The allowed values are the following.

INFORMATIONAL - No issue was found.

LOW - The issue does not require action on its own.

MEDIUM - The issue must be addressed but not urgently.

HIGH - The issue must be addressed as a priority.

CRITICAL - The issue must be remediated immediately to avoid it escalating.

If you provide Normalized and do not provide Label, then Label is set automatically as follows.

0 - INFORMATIONAL

1–39 - LOW

40–69 - MEDIUM

70–89 - HIGH

90–100 - CRITICAL

severity_original :: Lens' Severity (Maybe Text)

The native severity from the finding product that generated the finding.

severity_normalized :: Lens' Severity (Maybe Int)

Deprecated. The normalized severity of a finding. This attribute is being deprecated. Instead of providing Normalized, provide Label.

If you provide Label and do not provide Normalized, then Normalized is set automatically as follows.

INFORMATIONAL - 0

LOW - 1

MEDIUM - 40

HIGH - 70

CRITICAL - 90