Network.AWS.SecurityHub.Types.StandardsSubscriptionRequest

Description

Documentation

data StandardsSubscriptionRequest

The standard that you want to enable.

See: newStandardsSubscriptionRequest smart constructor.

Constructors

StandardsSubscriptionRequest'

Fields

standardsInput :: Maybe (HashMap Text Text)

A key-value pair of input for the standard.

standardsArn :: Text

The ARN of the standard that you want to enable. To view the list of available standards and their ARNs, use the DescribeStandards operation.

newStandardsSubscriptionRequest

Arguments

:: Text
$sel:standardsArn:StandardsSubscriptionRequest'

-> StandardsSubscriptionRequest

Create a value of StandardsSubscriptionRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:standardsInput:StandardsSubscriptionRequest', standardsSubscriptionRequest_standardsInput - A key-value pair of input for the standard.

$sel:standardsArn:StandardsSubscriptionRequest', standardsSubscriptionRequest_standardsArn - The ARN of the standard that you want to enable. To view the list of available standards and their ARNs, use the DescribeStandards operation.

standardsSubscriptionRequest_standardsInput :: Lens' StandardsSubscriptionRequest (Maybe (HashMap Text Text))

A key-value pair of input for the standard.

standardsSubscriptionRequest_standardsArn :: Lens' StandardsSubscriptionRequest Text

The ARN of the standard that you want to enable. To view the list of available standards and their ARNs, use the DescribeStandards operation.