Network.AWS.SecurityHub.Types.ThreatIntelIndicator

Description

Documentation

data ThreatIntelIndicator

Details about the threat intelligence related to a finding.

See: newThreatIntelIndicator smart constructor.

Constructors

ThreatIntelIndicator'

Fields

category :: Maybe ThreatIntelIndicatorCategory

The category of a threat intelligence indicator.

value :: Maybe Text

The value of a threat intelligence indicator.

source :: Maybe Text

The source of the threat intelligence indicator.

type' :: Maybe ThreatIntelIndicatorType

The type of threat intelligence indicator.

sourceUrl :: Maybe Text

The URL to the page or site where you can get more information about the threat intelligence indicator.

lastObservedAt :: Maybe Text

Indicates when the most recent instance of a threat intelligence indicator was observed.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

newThreatIntelIndicator :: ThreatIntelIndicator

Create a value of ThreatIntelIndicator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:category:ThreatIntelIndicator', threatIntelIndicator_category - The category of a threat intelligence indicator.

$sel:value:ThreatIntelIndicator', threatIntelIndicator_value - The value of a threat intelligence indicator.

$sel:source:ThreatIntelIndicator', threatIntelIndicator_source - The source of the threat intelligence indicator.

$sel:type':ThreatIntelIndicator', threatIntelIndicator_type - The type of threat intelligence indicator.

$sel:sourceUrl:ThreatIntelIndicator', threatIntelIndicator_sourceUrl - The URL to the page or site where you can get more information about the threat intelligence indicator.

$sel:lastObservedAt:ThreatIntelIndicator', threatIntelIndicator_lastObservedAt - Indicates when the most recent instance of a threat intelligence indicator was observed.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

threatIntelIndicator_category :: Lens' ThreatIntelIndicator (Maybe ThreatIntelIndicatorCategory)

The category of a threat intelligence indicator.

threatIntelIndicator_value :: Lens' ThreatIntelIndicator (Maybe Text)

The value of a threat intelligence indicator.

threatIntelIndicator_source :: Lens' ThreatIntelIndicator (Maybe Text)

The source of the threat intelligence indicator.

threatIntelIndicator_type :: Lens' ThreatIntelIndicator (Maybe ThreatIntelIndicatorType)

The type of threat intelligence indicator.

threatIntelIndicator_sourceUrl :: Lens' ThreatIntelIndicator (Maybe Text)

The URL to the page or site where you can get more information about the threat intelligence indicator.

threatIntelIndicator_lastObservedAt :: Lens' ThreatIntelIndicator (Maybe Text)

Indicates when the most recent instance of a threat intelligence indicator was observed.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.