Network.AWS.SecurityHub.Types.Vulnerability

Description

Documentation

data Vulnerability

A vulnerability associated with a finding.

See: newVulnerability smart constructor.

Constructors

Vulnerability'

Fields

vendor :: Maybe VulnerabilityVendor

Information about the vendor that generates the vulnerability report.

relatedVulnerabilities :: Maybe [Text]

List of vulnerabilities that are related to this vulnerability.

vulnerablePackages :: Maybe [SoftwarePackage]

List of software packages that have the vulnerability.

referenceUrls :: Maybe [Text]

A list of URLs that provide additional information about the vulnerability.

cvss :: Maybe [Cvss]

CVSS scores from the advisory related to the vulnerability.

id :: Text

The identifier of the vulnerability.

newVulnerability

Arguments

:: Text
$sel:id:Vulnerability'

-> Vulnerability

Create a value of Vulnerability with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vendor:Vulnerability', vulnerability_vendor - Information about the vendor that generates the vulnerability report.

$sel:relatedVulnerabilities:Vulnerability', vulnerability_relatedVulnerabilities - List of vulnerabilities that are related to this vulnerability.

$sel:vulnerablePackages:Vulnerability', vulnerability_vulnerablePackages - List of software packages that have the vulnerability.

$sel:referenceUrls:Vulnerability', vulnerability_referenceUrls - A list of URLs that provide additional information about the vulnerability.

$sel:cvss:Vulnerability', vulnerability_cvss - CVSS scores from the advisory related to the vulnerability.

$sel:id:Vulnerability', vulnerability_id - The identifier of the vulnerability.

vulnerability_vendor :: Lens' Vulnerability (Maybe VulnerabilityVendor)

Information about the vendor that generates the vulnerability report.

vulnerability_relatedVulnerabilities :: Lens' Vulnerability (Maybe [Text])

List of vulnerabilities that are related to this vulnerability.

vulnerability_vulnerablePackages :: Lens' Vulnerability (Maybe [SoftwarePackage])

List of software packages that have the vulnerability.

vulnerability_referenceUrls :: Lens' Vulnerability (Maybe [Text])

A list of URLs that provide additional information about the vulnerability.

vulnerability_cvss :: Lens' Vulnerability (Maybe [Cvss])

CVSS scores from the advisory related to the vulnerability.

vulnerability_id :: Lens' Vulnerability Text

The identifier of the vulnerability.