Network.AWS.SecurityHub.UpdateFindings

Description

UpdateFindings is deprecated. Instead of UpdateFindings, use BatchUpdateFindings.

Updates the Note and RecordState of the Security Hub-aggregated findings that the filter attributes specify. Any member account that can view the finding also sees the update to the finding.

Creating a Request

data UpdateFindings

See: newUpdateFindings smart constructor.

Constructors

UpdateFindings'

Fields

recordState :: Maybe RecordState

The updated record state for the finding.

note :: Maybe NoteUpdate

The updated note for the finding.

filters :: AwsSecurityFindingFilters

A collection of attributes that specify which findings you want to update.

newUpdateFindings

Arguments

:: AwsSecurityFindingFilters
$sel:filters:UpdateFindings'

-> UpdateFindings

Create a value of UpdateFindings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recordState:UpdateFindings', updateFindings_recordState - The updated record state for the finding.

$sel:note:UpdateFindings', updateFindings_note - The updated note for the finding.

$sel:filters:UpdateFindings', updateFindings_filters - A collection of attributes that specify which findings you want to update.

Request Lenses

updateFindings_recordState :: Lens' UpdateFindings (Maybe RecordState)

The updated record state for the finding.

updateFindings_note :: Lens' UpdateFindings (Maybe NoteUpdate)

The updated note for the finding.

updateFindings_filters :: Lens' UpdateFindings AwsSecurityFindingFilters

A collection of attributes that specify which findings you want to update.

Destructuring the Response

data UpdateFindingsResponse

See: newUpdateFindingsResponse smart constructor.

Constructors

UpdateFindingsResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateFindingsResponse

Arguments

:: Int
$sel:httpStatus:UpdateFindingsResponse'

-> UpdateFindingsResponse

Create a value of UpdateFindingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateFindingsResponse', updateFindingsResponse_httpStatus - The response's http status code.

Response Lenses

updateFindingsResponse_httpStatus :: Lens' UpdateFindingsResponse Int

The response's http status code.