Network.AWS.SecurityHub.UpdateSecurityHubConfiguration

Description

Updates configuration options for Security Hub.

Creating a Request

data UpdateSecurityHubConfiguration

See: newUpdateSecurityHubConfiguration smart constructor.

Constructors

UpdateSecurityHubConfiguration'

Fields

autoEnableControls :: Maybe Bool

Whether to automatically enable new controls when they are added to standards that are enabled.

By default, this is set to true, and new controls are enabled automatically. To not automatically enable new controls, set this to false.

newUpdateSecurityHubConfiguration :: UpdateSecurityHubConfiguration

Create a value of UpdateSecurityHubConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoEnableControls:UpdateSecurityHubConfiguration', updateSecurityHubConfiguration_autoEnableControls - Whether to automatically enable new controls when they are added to standards that are enabled.

By default, this is set to true, and new controls are enabled automatically. To not automatically enable new controls, set this to false.

Request Lenses

updateSecurityHubConfiguration_autoEnableControls :: Lens' UpdateSecurityHubConfiguration (Maybe Bool)

Whether to automatically enable new controls when they are added to standards that are enabled.

By default, this is set to true, and new controls are enabled automatically. To not automatically enable new controls, set this to false.

Destructuring the Response

data UpdateSecurityHubConfigurationResponse

See: newUpdateSecurityHubConfigurationResponse smart constructor.

Constructors

UpdateSecurityHubConfigurationResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateSecurityHubConfigurationResponse

Arguments

:: Int
$sel:httpStatus:UpdateSecurityHubConfigurationResponse'

-> UpdateSecurityHubConfigurationResponse

Create a value of UpdateSecurityHubConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateSecurityHubConfigurationResponse', updateSecurityHubConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

updateSecurityHubConfigurationResponse_httpStatus :: Lens' UpdateSecurityHubConfigurationResponse Int

The response's http status code.