Network.AWS.SecurityHub.UpdateStandardsControl

Description

Used to control whether an individual security standard control is enabled or disabled.

Creating a Request

data UpdateStandardsControl

See: newUpdateStandardsControl smart constructor.

Constructors

UpdateStandardsControl'

Fields

controlStatus :: Maybe ControlStatus

The updated status of the security standard control.

disabledReason :: Maybe Text

A description of the reason why you are disabling a security standard control. If you are disabling a control, then this is required.

standardsControlArn :: Text

The ARN of the security standard control to enable or disable.

newUpdateStandardsControl

Arguments

:: Text
$sel:standardsControlArn:UpdateStandardsControl'

-> UpdateStandardsControl

Create a value of UpdateStandardsControl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:controlStatus:UpdateStandardsControl', updateStandardsControl_controlStatus - The updated status of the security standard control.

$sel:disabledReason:UpdateStandardsControl', updateStandardsControl_disabledReason - A description of the reason why you are disabling a security standard control. If you are disabling a control, then this is required.

$sel:standardsControlArn:UpdateStandardsControl', updateStandardsControl_standardsControlArn - The ARN of the security standard control to enable or disable.

Request Lenses

updateStandardsControl_controlStatus :: Lens' UpdateStandardsControl (Maybe ControlStatus)

The updated status of the security standard control.

updateStandardsControl_disabledReason :: Lens' UpdateStandardsControl (Maybe Text)

A description of the reason why you are disabling a security standard control. If you are disabling a control, then this is required.

updateStandardsControl_standardsControlArn :: Lens' UpdateStandardsControl Text

The ARN of the security standard control to enable or disable.

Destructuring the Response

data UpdateStandardsControlResponse

See: newUpdateStandardsControlResponse smart constructor.

Constructors

UpdateStandardsControlResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateStandardsControlResponse

Arguments

:: Int
$sel:httpStatus:UpdateStandardsControlResponse'

-> UpdateStandardsControlResponse

Create a value of UpdateStandardsControlResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateStandardsControlResponse', updateStandardsControlResponse_httpStatus - The response's http status code.

Response Lenses

updateStandardsControlResponse_httpStatus :: Lens' UpdateStandardsControlResponse Int

The response's http status code.