Network.AWS.ServerlessApplicationRepository.CreateApplication

Description

Creates an application, optionally including an AWS SAM file to create the first application version in the same call.

Creating a Request

data CreateApplication

See: newCreateApplication smart constructor.

Constructors

CreateApplication'

Fields

spdxLicenseId :: Maybe Text

A valid identifier from https://spdx.org/licenses/.

licenseBody :: Maybe Text

A local text file that contains the license of the app that matches the spdxLicenseID value of your application. The file has the format file://<path>/<filename>.

Maximum size 5 MB

You can specify only one of licenseBody and licenseUrl; otherwise, an error results.

templateUrl :: Maybe Text

A link to the S3 object containing the packaged AWS SAM template of your application.

You can specify only one of templateBody and templateUrl; otherwise an error results.

licenseUrl :: Maybe Text

A link to the S3 object that contains the license of the app that matches the spdxLicenseID value of your application.

Maximum size 5 MB

You can specify only one of licenseBody and licenseUrl; otherwise, an error results.

labels :: Maybe [Text]

Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

readmeBody :: Maybe Text

A local text readme file in Markdown language that contains a more detailed description of the application and how it works. The file has the format file://<path>/<filename>.

Maximum size 5 MB

You can specify only one of readmeBody and readmeUrl; otherwise, an error results.

homePageUrl :: Maybe Text

A URL with more information about the application, for example the location of your GitHub repository for the application.

sourceCodeArchiveUrl :: Maybe Text

A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

readmeUrl :: Maybe Text

A link to the S3 object in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

You can specify only one of readmeBody and readmeUrl; otherwise, an error results.

sourceCodeUrl :: Maybe Text

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

templateBody :: Maybe Text

The local raw packaged AWS SAM template file of your application. The file has the format file://<path>/<filename>.

You can specify only one of templateBody and templateUrl; otherwise an error results.

semanticVersion :: Maybe Text

The semantic version of the application:

https://semver.org/

description :: Text

The description of the application.

Minimum length=1. Maximum length=256

name :: Text

The name of the application that you want to publish.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

author :: Text

The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

newCreateApplication

Arguments

:: Text
$sel:description:CreateApplication'

-> Text
$sel:name:CreateApplication'

-> Text
$sel:author:CreateApplication'

-> CreateApplication

Create a value of CreateApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:spdxLicenseId:CreateApplication', createApplication_spdxLicenseId - A valid identifier from https://spdx.org/licenses/.

$sel:licenseBody:CreateApplication', createApplication_licenseBody - A local text file that contains the license of the app that matches the spdxLicenseID value of your application. The file has the format file://<path>/<filename>.

Maximum size 5 MB

You can specify only one of licenseBody and licenseUrl; otherwise, an error results.

$sel:templateUrl:CreateApplication', createApplication_templateUrl - A link to the S3 object containing the packaged AWS SAM template of your application.

You can specify only one of templateBody and templateUrl; otherwise an error results.

$sel:licenseUrl:CreateApplication', createApplication_licenseUrl - A link to the S3 object that contains the license of the app that matches the spdxLicenseID value of your application.

Maximum size 5 MB

You can specify only one of licenseBody and licenseUrl; otherwise, an error results.

$sel:labels:CreateApplication', createApplication_labels - Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

$sel:readmeBody:CreateApplication', createApplication_readmeBody - A local text readme file in Markdown language that contains a more detailed description of the application and how it works. The file has the format file://<path>/<filename>.

Maximum size 5 MB

You can specify only one of readmeBody and readmeUrl; otherwise, an error results.

$sel:homePageUrl:CreateApplication', createApplication_homePageUrl - A URL with more information about the application, for example the location of your GitHub repository for the application.

$sel:sourceCodeArchiveUrl:CreateApplication', createApplication_sourceCodeArchiveUrl - A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

$sel:readmeUrl:CreateApplication', createApplication_readmeUrl - A link to the S3 object in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

You can specify only one of readmeBody and readmeUrl; otherwise, an error results.

$sel:sourceCodeUrl:CreateApplication', createApplication_sourceCodeUrl - A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

$sel:templateBody:CreateApplication', createApplication_templateBody - The local raw packaged AWS SAM template file of your application. The file has the format file://<path>/<filename>.

You can specify only one of templateBody and templateUrl; otherwise an error results.

$sel:semanticVersion:CreateApplication', createApplication_semanticVersion - The semantic version of the application:

https://semver.org/

$sel:description:CreateApplication', createApplication_description - The description of the application.

Minimum length=1. Maximum length=256

$sel:name:CreateApplication', createApplication_name - The name of the application that you want to publish.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

$sel:author:CreateApplication', createApplication_author - The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

Request Lenses

createApplication_spdxLicenseId :: Lens' CreateApplication (Maybe Text)

A valid identifier from https://spdx.org/licenses/.

createApplication_licenseBody :: Lens' CreateApplication (Maybe Text)

A local text file that contains the license of the app that matches the spdxLicenseID value of your application. The file has the format file://<path>/<filename>.

Maximum size 5 MB

You can specify only one of licenseBody and licenseUrl; otherwise, an error results.

createApplication_templateUrl :: Lens' CreateApplication (Maybe Text)

A link to the S3 object containing the packaged AWS SAM template of your application.

You can specify only one of templateBody and templateUrl; otherwise an error results.

createApplication_licenseUrl :: Lens' CreateApplication (Maybe Text)

A link to the S3 object that contains the license of the app that matches the spdxLicenseID value of your application.

Maximum size 5 MB

You can specify only one of licenseBody and licenseUrl; otherwise, an error results.

createApplication_labels :: Lens' CreateApplication (Maybe [Text])

Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

createApplication_readmeBody :: Lens' CreateApplication (Maybe Text)

A local text readme file in Markdown language that contains a more detailed description of the application and how it works. The file has the format file://<path>/<filename>.

Maximum size 5 MB

You can specify only one of readmeBody and readmeUrl; otherwise, an error results.

createApplication_homePageUrl :: Lens' CreateApplication (Maybe Text)

A URL with more information about the application, for example the location of your GitHub repository for the application.

createApplication_sourceCodeArchiveUrl :: Lens' CreateApplication (Maybe Text)

A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

createApplication_readmeUrl :: Lens' CreateApplication (Maybe Text)

A link to the S3 object in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

You can specify only one of readmeBody and readmeUrl; otherwise, an error results.

createApplication_sourceCodeUrl :: Lens' CreateApplication (Maybe Text)

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

createApplication_templateBody :: Lens' CreateApplication (Maybe Text)

The local raw packaged AWS SAM template file of your application. The file has the format file://<path>/<filename>.

You can specify only one of templateBody and templateUrl; otherwise an error results.

createApplication_semanticVersion :: Lens' CreateApplication (Maybe Text)

The semantic version of the application:

https://semver.org/

createApplication_description :: Lens' CreateApplication Text

The description of the application.

Minimum length=1. Maximum length=256

createApplication_name :: Lens' CreateApplication Text

The name of the application that you want to publish.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

createApplication_author :: Lens' CreateApplication Text

The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

Destructuring the Response

data CreateApplicationResponse

See: newCreateApplicationResponse smart constructor.

Constructors

CreateApplicationResponse'

Fields

applicationId :: Maybe Text

The application Amazon Resource Name (ARN).

creationTime :: Maybe Text

The date and time this resource was created.

spdxLicenseId :: Maybe Text

A valid identifier from https://spdx.org/licenses/.

licenseUrl :: Maybe Text

A link to a license file of the app that matches the spdxLicenseID value of your application.

Maximum size 5 MB

verifiedAuthorUrl :: Maybe Text

The URL to the public profile of a verified author. This URL is submitted by the author.

labels :: Maybe [Text]

Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

author :: Maybe Text

The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

version :: Maybe Version

Version information about the application.

name :: Maybe Text

The name of the application.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

homePageUrl :: Maybe Text

A URL with more information about the application, for example the location of your GitHub repository for the application.

readmeUrl :: Maybe Text

A link to the readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

isVerifiedAuthor :: Maybe Bool

Whether the author of this application has been verified. This means means that AWS has made a good faith review, as a reasonable and prudent service provider, of the information provided by the requester and has confirmed that the requester's identity is as claimed.

description :: Maybe Text

The description of the application.

Minimum length=1. Maximum length=256

httpStatus :: Int

The response's http status code.

newCreateApplicationResponse

Arguments

:: Int
$sel:httpStatus:CreateApplicationResponse'

-> CreateApplicationResponse

Create a value of CreateApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:CreateApplicationResponse', createApplicationResponse_applicationId - The application Amazon Resource Name (ARN).

$sel:creationTime:CreateApplicationResponse', createApplicationResponse_creationTime - The date and time this resource was created.

$sel:spdxLicenseId:CreateApplication', createApplicationResponse_spdxLicenseId - A valid identifier from https://spdx.org/licenses/.

$sel:licenseUrl:CreateApplication', createApplicationResponse_licenseUrl - A link to a license file of the app that matches the spdxLicenseID value of your application.

Maximum size 5 MB

$sel:verifiedAuthorUrl:CreateApplicationResponse', createApplicationResponse_verifiedAuthorUrl - The URL to the public profile of a verified author. This URL is submitted by the author.

$sel:labels:CreateApplication', createApplicationResponse_labels - Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

$sel:author:CreateApplication', createApplicationResponse_author - The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

$sel:version:CreateApplicationResponse', createApplicationResponse_version - Version information about the application.

$sel:name:CreateApplication', createApplicationResponse_name - The name of the application.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

$sel:homePageUrl:CreateApplication', createApplicationResponse_homePageUrl - A URL with more information about the application, for example the location of your GitHub repository for the application.

$sel:readmeUrl:CreateApplication', createApplicationResponse_readmeUrl - A link to the readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

$sel:isVerifiedAuthor:CreateApplicationResponse', createApplicationResponse_isVerifiedAuthor - Whether the author of this application has been verified. This means means that AWS has made a good faith review, as a reasonable and prudent service provider, of the information provided by the requester and has confirmed that the requester's identity is as claimed.

$sel:description:CreateApplication', createApplicationResponse_description - The description of the application.

Minimum length=1. Maximum length=256

$sel:httpStatus:CreateApplicationResponse', createApplicationResponse_httpStatus - The response's http status code.

Response Lenses

createApplicationResponse_applicationId :: Lens' CreateApplicationResponse (Maybe Text)

The application Amazon Resource Name (ARN).

createApplicationResponse_creationTime :: Lens' CreateApplicationResponse (Maybe Text)

The date and time this resource was created.

createApplicationResponse_spdxLicenseId :: Lens' CreateApplicationResponse (Maybe Text)

A valid identifier from https://spdx.org/licenses/.

createApplicationResponse_licenseUrl :: Lens' CreateApplicationResponse (Maybe Text)

A link to a license file of the app that matches the spdxLicenseID value of your application.

Maximum size 5 MB

createApplicationResponse_verifiedAuthorUrl :: Lens' CreateApplicationResponse (Maybe Text)

The URL to the public profile of a verified author. This URL is submitted by the author.

createApplicationResponse_labels :: Lens' CreateApplicationResponse (Maybe [Text])

Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

createApplicationResponse_author :: Lens' CreateApplicationResponse (Maybe Text)

The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

createApplicationResponse_version :: Lens' CreateApplicationResponse (Maybe Version)

Version information about the application.

createApplicationResponse_name :: Lens' CreateApplicationResponse (Maybe Text)

The name of the application.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

createApplicationResponse_homePageUrl :: Lens' CreateApplicationResponse (Maybe Text)

A URL with more information about the application, for example the location of your GitHub repository for the application.

createApplicationResponse_readmeUrl :: Lens' CreateApplicationResponse (Maybe Text)

A link to the readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

createApplicationResponse_isVerifiedAuthor :: Lens' CreateApplicationResponse (Maybe Bool)

Whether the author of this application has been verified. This means means that AWS has made a good faith review, as a reasonable and prudent service provider, of the information provided by the requester and has confirmed that the requester's identity is as claimed.

createApplicationResponse_description :: Lens' CreateApplicationResponse (Maybe Text)

The description of the application.

Minimum length=1. Maximum length=256

createApplicationResponse_httpStatus :: Lens' CreateApplicationResponse Int

The response's http status code.