Network.AWS.ServerlessApplicationRepository.CreateCloudFormationChangeSet

Description

Creates an AWS CloudFormation change set for the given application.

Creating a Request

data CreateCloudFormationChangeSet

See: newCreateCloudFormationChangeSet smart constructor.

Constructors

CreateCloudFormationChangeSet'

Fields

resourceTypes :: Maybe [Text]

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

capabilities :: Maybe [Text]

A list of values that you must specify before you can deploy certain applications. Some applications might include resources that can affect permissions in your AWS account, for example, by creating new AWS Identity and Access Management (IAM) users. For those applications, you must explicitly acknowledge their capabilities by specifying this parameter.

The only valid values are CAPABILITY_IAM, CAPABILITY_NAMED_IAM, CAPABILITY_RESOURCE_POLICY, and CAPABILITY_AUTO_EXPAND.

The following resources require you to specify CAPABILITY_IAM or CAPABILITY_NAMED_IAM: AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, and AWS::IAM::Role. If the application contains IAM resources, you can specify either CAPABILITY_IAM or CAPABILITY_NAMED_IAM. If the application contains IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

The following resources require you to specify CAPABILITY_RESOURCE_POLICY: AWS::Lambda::Permission, AWS::IAM:Policy, AWS::ApplicationAutoScaling::ScalingPolicy, AWS::S3::BucketPolicy, AWS::SQS::QueuePolicy, and AWS::SNS:TopicPolicy.

Applications that contain one or more nested applications require you to specify CAPABILITY_AUTO_EXPAND.

If your application template contains any of the above resources, we recommend that you review all permissions associated with the application before deploying. If you don't specify this parameter for an application that requires capabilities, the call will fail.

parameterOverrides :: Maybe [ParameterValue]

A list of parameter values for the parameters of the application.

notificationArns :: Maybe [Text]

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

tags :: Maybe [Tag]

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

rollbackConfiguration :: Maybe RollbackConfiguration

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

description :: Maybe Text

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

changeSetName :: Maybe Text

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

semanticVersion :: Maybe Text

The semantic version of the application:

https://semver.org/

templateId :: Maybe Text

The UUID returned by CreateCloudFormationTemplate.

Pattern: [0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}

clientToken :: Maybe Text

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

applicationId :: Text

The Amazon Resource Name (ARN) of the application.

stackName :: Text

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

newCreateCloudFormationChangeSet

Arguments

:: Text
$sel:applicationId:CreateCloudFormationChangeSet'

-> Text
$sel:stackName:CreateCloudFormationChangeSet'

-> CreateCloudFormationChangeSet

Create a value of CreateCloudFormationChangeSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceTypes:CreateCloudFormationChangeSet', createCloudFormationChangeSet_resourceTypes - This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

$sel:capabilities:CreateCloudFormationChangeSet', createCloudFormationChangeSet_capabilities - A list of values that you must specify before you can deploy certain applications. Some applications might include resources that can affect permissions in your AWS account, for example, by creating new AWS Identity and Access Management (IAM) users. For those applications, you must explicitly acknowledge their capabilities by specifying this parameter.

The only valid values are CAPABILITY_IAM, CAPABILITY_NAMED_IAM, CAPABILITY_RESOURCE_POLICY, and CAPABILITY_AUTO_EXPAND.

The following resources require you to specify CAPABILITY_IAM or CAPABILITY_NAMED_IAM: AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, and AWS::IAM::Role. If the application contains IAM resources, you can specify either CAPABILITY_IAM or CAPABILITY_NAMED_IAM. If the application contains IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

The following resources require you to specify CAPABILITY_RESOURCE_POLICY: AWS::Lambda::Permission, AWS::IAM:Policy, AWS::ApplicationAutoScaling::ScalingPolicy, AWS::S3::BucketPolicy, AWS::SQS::QueuePolicy, and AWS::SNS:TopicPolicy.

Applications that contain one or more nested applications require you to specify CAPABILITY_AUTO_EXPAND.

If your application template contains any of the above resources, we recommend that you review all permissions associated with the application before deploying. If you don't specify this parameter for an application that requires capabilities, the call will fail.

$sel:parameterOverrides:CreateCloudFormationChangeSet', createCloudFormationChangeSet_parameterOverrides - A list of parameter values for the parameters of the application.

$sel:notificationArns:CreateCloudFormationChangeSet', createCloudFormationChangeSet_notificationArns - This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

$sel:tags:CreateCloudFormationChangeSet', createCloudFormationChangeSet_tags - This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

$sel:rollbackConfiguration:CreateCloudFormationChangeSet', createCloudFormationChangeSet_rollbackConfiguration - This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

$sel:description:CreateCloudFormationChangeSet', createCloudFormationChangeSet_description - This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

$sel:changeSetName:CreateCloudFormationChangeSet', createCloudFormationChangeSet_changeSetName - This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

$sel:semanticVersion:CreateCloudFormationChangeSet', createCloudFormationChangeSet_semanticVersion - The semantic version of the application:

https://semver.org/

$sel:templateId:CreateCloudFormationChangeSet', createCloudFormationChangeSet_templateId - The UUID returned by CreateCloudFormationTemplate.

Pattern: [0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}

$sel:clientToken:CreateCloudFormationChangeSet', createCloudFormationChangeSet_clientToken - This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

$sel:applicationId:CreateCloudFormationChangeSet', createCloudFormationChangeSet_applicationId - The Amazon Resource Name (ARN) of the application.

$sel:stackName:CreateCloudFormationChangeSet', createCloudFormationChangeSet_stackName - This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

Request Lenses

createCloudFormationChangeSet_resourceTypes :: Lens' CreateCloudFormationChangeSet (Maybe [Text])

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

createCloudFormationChangeSet_capabilities :: Lens' CreateCloudFormationChangeSet (Maybe [Text])

A list of values that you must specify before you can deploy certain applications. Some applications might include resources that can affect permissions in your AWS account, for example, by creating new AWS Identity and Access Management (IAM) users. For those applications, you must explicitly acknowledge their capabilities by specifying this parameter.

The only valid values are CAPABILITY_IAM, CAPABILITY_NAMED_IAM, CAPABILITY_RESOURCE_POLICY, and CAPABILITY_AUTO_EXPAND.

The following resources require you to specify CAPABILITY_IAM or CAPABILITY_NAMED_IAM: AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, and AWS::IAM::Role. If the application contains IAM resources, you can specify either CAPABILITY_IAM or CAPABILITY_NAMED_IAM. If the application contains IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

The following resources require you to specify CAPABILITY_RESOURCE_POLICY: AWS::Lambda::Permission, AWS::IAM:Policy, AWS::ApplicationAutoScaling::ScalingPolicy, AWS::S3::BucketPolicy, AWS::SQS::QueuePolicy, and AWS::SNS:TopicPolicy.

Applications that contain one or more nested applications require you to specify CAPABILITY_AUTO_EXPAND.

If your application template contains any of the above resources, we recommend that you review all permissions associated with the application before deploying. If you don't specify this parameter for an application that requires capabilities, the call will fail.

createCloudFormationChangeSet_parameterOverrides :: Lens' CreateCloudFormationChangeSet (Maybe [ParameterValue])

A list of parameter values for the parameters of the application.

createCloudFormationChangeSet_notificationArns :: Lens' CreateCloudFormationChangeSet (Maybe [Text])

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

createCloudFormationChangeSet_tags :: Lens' CreateCloudFormationChangeSet (Maybe [Tag])

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

createCloudFormationChangeSet_rollbackConfiguration :: Lens' CreateCloudFormationChangeSet (Maybe RollbackConfiguration)

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

createCloudFormationChangeSet_description :: Lens' CreateCloudFormationChangeSet (Maybe Text)

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

createCloudFormationChangeSet_changeSetName :: Lens' CreateCloudFormationChangeSet (Maybe Text)

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

createCloudFormationChangeSet_semanticVersion :: Lens' CreateCloudFormationChangeSet (Maybe Text)

The semantic version of the application:

https://semver.org/

createCloudFormationChangeSet_templateId :: Lens' CreateCloudFormationChangeSet (Maybe Text)

The UUID returned by CreateCloudFormationTemplate.

Pattern: [0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}

createCloudFormationChangeSet_clientToken :: Lens' CreateCloudFormationChangeSet (Maybe Text)

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

createCloudFormationChangeSet_applicationId :: Lens' CreateCloudFormationChangeSet Text

The Amazon Resource Name (ARN) of the application.

createCloudFormationChangeSet_stackName :: Lens' CreateCloudFormationChangeSet Text

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

Destructuring the Response

data CreateCloudFormationChangeSetResponse

See: newCreateCloudFormationChangeSetResponse smart constructor.

Constructors

CreateCloudFormationChangeSetResponse'

Fields

applicationId :: Maybe Text

The application Amazon Resource Name (ARN).

stackId :: Maybe Text

The unique ID of the stack.

changeSetId :: Maybe Text

The Amazon Resource Name (ARN) of the change set.

Length constraints: Minimum length of 1.

Pattern: ARN:[-a-zA-Z0-9:/]*

semanticVersion :: Maybe Text

The semantic version of the application:

https://semver.org/

httpStatus :: Int

The response's http status code.

newCreateCloudFormationChangeSetResponse

Arguments

:: Int
$sel:httpStatus:CreateCloudFormationChangeSetResponse'

-> CreateCloudFormationChangeSetResponse

Create a value of CreateCloudFormationChangeSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:CreateCloudFormationChangeSet', createCloudFormationChangeSetResponse_applicationId - The application Amazon Resource Name (ARN).

$sel:stackId:CreateCloudFormationChangeSetResponse', createCloudFormationChangeSetResponse_stackId - The unique ID of the stack.

$sel:changeSetId:CreateCloudFormationChangeSetResponse', createCloudFormationChangeSetResponse_changeSetId - The Amazon Resource Name (ARN) of the change set.

Length constraints: Minimum length of 1.

Pattern: ARN:[-a-zA-Z0-9:/]*

$sel:semanticVersion:CreateCloudFormationChangeSet', createCloudFormationChangeSetResponse_semanticVersion - The semantic version of the application:

https://semver.org/

$sel:httpStatus:CreateCloudFormationChangeSetResponse', createCloudFormationChangeSetResponse_httpStatus - The response's http status code.

Response Lenses

createCloudFormationChangeSetResponse_applicationId :: Lens' CreateCloudFormationChangeSetResponse (Maybe Text)

The application Amazon Resource Name (ARN).

createCloudFormationChangeSetResponse_stackId :: Lens' CreateCloudFormationChangeSetResponse (Maybe Text)

The unique ID of the stack.

createCloudFormationChangeSetResponse_changeSetId :: Lens' CreateCloudFormationChangeSetResponse (Maybe Text)

The Amazon Resource Name (ARN) of the change set.

Length constraints: Minimum length of 1.

Pattern: ARN:[-a-zA-Z0-9:/]*

createCloudFormationChangeSetResponse_semanticVersion :: Lens' CreateCloudFormationChangeSetResponse (Maybe Text)

The semantic version of the application:

https://semver.org/

createCloudFormationChangeSetResponse_httpStatus :: Lens' CreateCloudFormationChangeSetResponse Int

The response's http status code.