Network.AWS.ServerlessApplicationRepository.GetApplication

Description

Gets the specified application.

Creating a Request

data GetApplication

See: newGetApplication smart constructor.

Constructors

GetApplication'

Fields

semanticVersion :: Maybe Text

The semantic version of the application to get.

applicationId :: Text

The Amazon Resource Name (ARN) of the application.

newGetApplication

Arguments

:: Text
$sel:applicationId:GetApplication'

-> GetApplication

Create a value of GetApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:semanticVersion:GetApplication', getApplication_semanticVersion - The semantic version of the application to get.

$sel:applicationId:GetApplication', getApplication_applicationId - The Amazon Resource Name (ARN) of the application.

Request Lenses

getApplication_semanticVersion :: Lens' GetApplication (Maybe Text)

The semantic version of the application to get.

getApplication_applicationId :: Lens' GetApplication Text

The Amazon Resource Name (ARN) of the application.

Destructuring the Response

data GetApplicationResponse

See: newGetApplicationResponse smart constructor.

Constructors

GetApplicationResponse'

Fields

applicationId :: Maybe Text

The application Amazon Resource Name (ARN).

creationTime :: Maybe Text

The date and time this resource was created.

spdxLicenseId :: Maybe Text

A valid identifier from https://spdx.org/licenses/.

licenseUrl :: Maybe Text

A link to a license file of the app that matches the spdxLicenseID value of your application.

Maximum size 5 MB

verifiedAuthorUrl :: Maybe Text

The URL to the public profile of a verified author. This URL is submitted by the author.

labels :: Maybe [Text]

Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

author :: Maybe Text

The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

version :: Maybe Version

Version information about the application.

name :: Maybe Text

The name of the application.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

homePageUrl :: Maybe Text

A URL with more information about the application, for example the location of your GitHub repository for the application.

readmeUrl :: Maybe Text

A link to the readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

isVerifiedAuthor :: Maybe Bool

Whether the author of this application has been verified. This means means that AWS has made a good faith review, as a reasonable and prudent service provider, of the information provided by the requester and has confirmed that the requester's identity is as claimed.

description :: Maybe Text

The description of the application.

Minimum length=1. Maximum length=256

httpStatus :: Int

The response's http status code.

newGetApplicationResponse

Arguments

:: Int
$sel:httpStatus:GetApplicationResponse'

-> GetApplicationResponse

Create a value of GetApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:GetApplication', getApplicationResponse_applicationId - The application Amazon Resource Name (ARN).

$sel:creationTime:GetApplicationResponse', getApplicationResponse_creationTime - The date and time this resource was created.

$sel:spdxLicenseId:GetApplicationResponse', getApplicationResponse_spdxLicenseId - A valid identifier from https://spdx.org/licenses/.

$sel:licenseUrl:GetApplicationResponse', getApplicationResponse_licenseUrl - A link to a license file of the app that matches the spdxLicenseID value of your application.

Maximum size 5 MB

$sel:verifiedAuthorUrl:GetApplicationResponse', getApplicationResponse_verifiedAuthorUrl - The URL to the public profile of a verified author. This URL is submitted by the author.

$sel:labels:GetApplicationResponse', getApplicationResponse_labels - Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

$sel:author:GetApplicationResponse', getApplicationResponse_author - The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

$sel:version:GetApplicationResponse', getApplicationResponse_version - Version information about the application.

$sel:name:GetApplicationResponse', getApplicationResponse_name - The name of the application.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

$sel:homePageUrl:GetApplicationResponse', getApplicationResponse_homePageUrl - A URL with more information about the application, for example the location of your GitHub repository for the application.

$sel:readmeUrl:GetApplicationResponse', getApplicationResponse_readmeUrl - A link to the readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

$sel:isVerifiedAuthor:GetApplicationResponse', getApplicationResponse_isVerifiedAuthor - Whether the author of this application has been verified. This means means that AWS has made a good faith review, as a reasonable and prudent service provider, of the information provided by the requester and has confirmed that the requester's identity is as claimed.

$sel:description:GetApplicationResponse', getApplicationResponse_description - The description of the application.

Minimum length=1. Maximum length=256

$sel:httpStatus:GetApplicationResponse', getApplicationResponse_httpStatus - The response's http status code.

Response Lenses

getApplicationResponse_applicationId :: Lens' GetApplicationResponse (Maybe Text)

The application Amazon Resource Name (ARN).

getApplicationResponse_creationTime :: Lens' GetApplicationResponse (Maybe Text)

The date and time this resource was created.

getApplicationResponse_spdxLicenseId :: Lens' GetApplicationResponse (Maybe Text)

A valid identifier from https://spdx.org/licenses/.

getApplicationResponse_licenseUrl :: Lens' GetApplicationResponse (Maybe Text)

A link to a license file of the app that matches the spdxLicenseID value of your application.

Maximum size 5 MB

getApplicationResponse_verifiedAuthorUrl :: Lens' GetApplicationResponse (Maybe Text)

The URL to the public profile of a verified author. This URL is submitted by the author.

getApplicationResponse_labels :: Lens' GetApplicationResponse (Maybe [Text])

Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

getApplicationResponse_author :: Lens' GetApplicationResponse (Maybe Text)

The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

getApplicationResponse_version :: Lens' GetApplicationResponse (Maybe Version)

Version information about the application.

getApplicationResponse_name :: Lens' GetApplicationResponse (Maybe Text)

The name of the application.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

getApplicationResponse_homePageUrl :: Lens' GetApplicationResponse (Maybe Text)

A URL with more information about the application, for example the location of your GitHub repository for the application.

getApplicationResponse_readmeUrl :: Lens' GetApplicationResponse (Maybe Text)

A link to the readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

getApplicationResponse_isVerifiedAuthor :: Lens' GetApplicationResponse (Maybe Bool)

Whether the author of this application has been verified. This means means that AWS has made a good faith review, as a reasonable and prudent service provider, of the information provided by the requester and has confirmed that the requester's identity is as claimed.

getApplicationResponse_description :: Lens' GetApplicationResponse (Maybe Text)

The description of the application.

Minimum length=1. Maximum length=256

getApplicationResponse_httpStatus :: Lens' GetApplicationResponse Int

The response's http status code.