Network.AWS.ServerlessApplicationRepository.Lens

Description

Operations

GetApplicationPolicy

getApplicationPolicy_applicationId :: Lens' GetApplicationPolicy Text

The Amazon Resource Name (ARN) of the application.

getApplicationPolicyResponse_statements :: Lens' GetApplicationPolicyResponse (Maybe [ApplicationPolicyStatement])

An array of policy statements applied to the application.

getApplicationPolicyResponse_httpStatus :: Lens' GetApplicationPolicyResponse Int

The response's http status code.

CreateApplicationVersion

createApplicationVersion_templateUrl :: Lens' CreateApplicationVersion (Maybe Text)

A link to the packaged AWS SAM template of your application.

createApplicationVersion_sourceCodeArchiveUrl :: Lens' CreateApplicationVersion (Maybe Text)

A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

createApplicationVersion_sourceCodeUrl :: Lens' CreateApplicationVersion (Maybe Text)

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

createApplicationVersion_templateBody :: Lens' CreateApplicationVersion (Maybe Text)

The raw packaged AWS SAM template of your application.

createApplicationVersion_applicationId :: Lens' CreateApplicationVersion Text

The Amazon Resource Name (ARN) of the application.

createApplicationVersion_semanticVersion :: Lens' CreateApplicationVersion Text

The semantic version of the new version.

createApplicationVersionResponse_applicationId :: Lens' CreateApplicationVersionResponse (Maybe Text)

The application Amazon Resource Name (ARN).

createApplicationVersionResponse_requiredCapabilities :: Lens' CreateApplicationVersionResponse (Maybe [Capability])

A list of values that you must specify before you can deploy certain applications. Some applications might include resources that can affect permissions in your AWS account, for example, by creating new AWS Identity and Access Management (IAM) users. For those applications, you must explicitly acknowledge their capabilities by specifying this parameter.

The only valid values are CAPABILITY_IAM, CAPABILITY_NAMED_IAM, CAPABILITY_RESOURCE_POLICY, and CAPABILITY_AUTO_EXPAND.

The following resources require you to specify CAPABILITY_IAM or CAPABILITY_NAMED_IAM: AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, and AWS::IAM::Role. If the application contains IAM resources, you can specify either CAPABILITY_IAM or CAPABILITY_NAMED_IAM. If the application contains IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

The following resources require you to specify CAPABILITY_RESOURCE_POLICY: AWS::Lambda::Permission, AWS::IAM:Policy, AWS::ApplicationAutoScaling::ScalingPolicy, AWS::S3::BucketPolicy, AWS::SQS::QueuePolicy, and AWS::SNS::TopicPolicy.

Applications that contain one or more nested applications require you to specify CAPABILITY_AUTO_EXPAND.

If your application template contains any of the above resources, we recommend that you review all permissions associated with the application before deploying. If you don't specify this parameter for an application that requires capabilities, the call will fail.

createApplicationVersionResponse_resourcesSupported :: Lens' CreateApplicationVersionResponse (Maybe Bool)

Whether all of the AWS resources contained in this application are supported in the region in which it is being retrieved.

createApplicationVersionResponse_parameterDefinitions :: Lens' CreateApplicationVersionResponse (Maybe [ParameterDefinition])

An array of parameter types supported by the application.

createApplicationVersionResponse_creationTime :: Lens' CreateApplicationVersionResponse (Maybe Text)

The date and time this resource was created.

createApplicationVersionResponse_templateUrl :: Lens' CreateApplicationVersionResponse (Maybe Text)

A link to the packaged AWS SAM template of your application.

createApplicationVersionResponse_sourceCodeArchiveUrl :: Lens' CreateApplicationVersionResponse (Maybe Text)

A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

createApplicationVersionResponse_sourceCodeUrl :: Lens' CreateApplicationVersionResponse (Maybe Text)

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

createApplicationVersionResponse_semanticVersion :: Lens' CreateApplicationVersionResponse (Maybe Text)

The semantic version of the application:

https://semver.org/

createApplicationVersionResponse_httpStatus :: Lens' CreateApplicationVersionResponse Int

The response's http status code.

GetCloudFormationTemplate

getCloudFormationTemplate_applicationId :: Lens' GetCloudFormationTemplate Text

The Amazon Resource Name (ARN) of the application.

getCloudFormationTemplate_templateId :: Lens' GetCloudFormationTemplate Text

The UUID returned by CreateCloudFormationTemplate.

Pattern: [0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}

getCloudFormationTemplateResponse_applicationId :: Lens' GetCloudFormationTemplateResponse (Maybe Text)

The application Amazon Resource Name (ARN).

getCloudFormationTemplateResponse_creationTime :: Lens' GetCloudFormationTemplateResponse (Maybe Text)

The date and time this resource was created.

getCloudFormationTemplateResponse_status :: Lens' GetCloudFormationTemplateResponse (Maybe Status)

Status of the template creation workflow.

Possible values: PREPARING | ACTIVE | EXPIRED

getCloudFormationTemplateResponse_expirationTime :: Lens' GetCloudFormationTemplateResponse (Maybe Text)

The date and time this template expires. Templates expire 1 hour after creation.

getCloudFormationTemplateResponse_templateUrl :: Lens' GetCloudFormationTemplateResponse (Maybe Text)

A link to the template that can be used to deploy the application using AWS CloudFormation.

getCloudFormationTemplateResponse_semanticVersion :: Lens' GetCloudFormationTemplateResponse (Maybe Text)

The semantic version of the application:

https://semver.org/

getCloudFormationTemplateResponse_templateId :: Lens' GetCloudFormationTemplateResponse (Maybe Text)

The UUID returned by CreateCloudFormationTemplate.

Pattern: [0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}

getCloudFormationTemplateResponse_httpStatus :: Lens' GetCloudFormationTemplateResponse Int

The response's http status code.

UnshareApplication

unshareApplication_applicationId :: Lens' UnshareApplication Text

The Amazon Resource Name (ARN) of the application.

unshareApplication_organizationId :: Lens' UnshareApplication Text

The AWS Organization ID to unshare the application from.

CreateApplication

createApplication_spdxLicenseId :: Lens' CreateApplication (Maybe Text)

A valid identifier from https://spdx.org/licenses/.

createApplication_licenseBody :: Lens' CreateApplication (Maybe Text)

A local text file that contains the license of the app that matches the spdxLicenseID value of your application. The file has the format file://<path>/<filename>.

Maximum size 5 MB

You can specify only one of licenseBody and licenseUrl; otherwise, an error results.

createApplication_templateUrl :: Lens' CreateApplication (Maybe Text)

A link to the S3 object containing the packaged AWS SAM template of your application.

You can specify only one of templateBody and templateUrl; otherwise an error results.

createApplication_licenseUrl :: Lens' CreateApplication (Maybe Text)

A link to the S3 object that contains the license of the app that matches the spdxLicenseID value of your application.

Maximum size 5 MB

You can specify only one of licenseBody and licenseUrl; otherwise, an error results.

createApplication_labels :: Lens' CreateApplication (Maybe [Text])

Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

createApplication_readmeBody :: Lens' CreateApplication (Maybe Text)

A local text readme file in Markdown language that contains a more detailed description of the application and how it works. The file has the format file://<path>/<filename>.

Maximum size 5 MB

You can specify only one of readmeBody and readmeUrl; otherwise, an error results.

createApplication_homePageUrl :: Lens' CreateApplication (Maybe Text)

A URL with more information about the application, for example the location of your GitHub repository for the application.

createApplication_sourceCodeArchiveUrl :: Lens' CreateApplication (Maybe Text)

A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

createApplication_readmeUrl :: Lens' CreateApplication (Maybe Text)

A link to the S3 object in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

You can specify only one of readmeBody and readmeUrl; otherwise, an error results.

createApplication_sourceCodeUrl :: Lens' CreateApplication (Maybe Text)

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

createApplication_templateBody :: Lens' CreateApplication (Maybe Text)

The local raw packaged AWS SAM template file of your application. The file has the format file://<path>/<filename>.

You can specify only one of templateBody and templateUrl; otherwise an error results.

createApplication_semanticVersion :: Lens' CreateApplication (Maybe Text)

The semantic version of the application:

https://semver.org/

createApplication_description :: Lens' CreateApplication Text

The description of the application.

Minimum length=1. Maximum length=256

createApplication_name :: Lens' CreateApplication Text

The name of the application that you want to publish.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

createApplication_author :: Lens' CreateApplication Text

The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

createApplicationResponse_applicationId :: Lens' CreateApplicationResponse (Maybe Text)

The application Amazon Resource Name (ARN).

createApplicationResponse_creationTime :: Lens' CreateApplicationResponse (Maybe Text)

The date and time this resource was created.

createApplicationResponse_spdxLicenseId :: Lens' CreateApplicationResponse (Maybe Text)

A valid identifier from https://spdx.org/licenses/.

createApplicationResponse_licenseUrl :: Lens' CreateApplicationResponse (Maybe Text)

A link to a license file of the app that matches the spdxLicenseID value of your application.

Maximum size 5 MB

createApplicationResponse_verifiedAuthorUrl :: Lens' CreateApplicationResponse (Maybe Text)

The URL to the public profile of a verified author. This URL is submitted by the author.

createApplicationResponse_labels :: Lens' CreateApplicationResponse (Maybe [Text])

Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

createApplicationResponse_author :: Lens' CreateApplicationResponse (Maybe Text)

The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

createApplicationResponse_version :: Lens' CreateApplicationResponse (Maybe Version)

Version information about the application.

createApplicationResponse_name :: Lens' CreateApplicationResponse (Maybe Text)

The name of the application.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

createApplicationResponse_homePageUrl :: Lens' CreateApplicationResponse (Maybe Text)

A URL with more information about the application, for example the location of your GitHub repository for the application.

createApplicationResponse_readmeUrl :: Lens' CreateApplicationResponse (Maybe Text)

A link to the readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

createApplicationResponse_isVerifiedAuthor :: Lens' CreateApplicationResponse (Maybe Bool)

Whether the author of this application has been verified. This means means that AWS has made a good faith review, as a reasonable and prudent service provider, of the information provided by the requester and has confirmed that the requester's identity is as claimed.

createApplicationResponse_description :: Lens' CreateApplicationResponse (Maybe Text)

The description of the application.

Minimum length=1. Maximum length=256

createApplicationResponse_httpStatus :: Lens' CreateApplicationResponse Int

The response's http status code.

CreateCloudFormationChangeSet

createCloudFormationChangeSet_resourceTypes :: Lens' CreateCloudFormationChangeSet (Maybe [Text])

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

createCloudFormationChangeSet_capabilities :: Lens' CreateCloudFormationChangeSet (Maybe [Text])

A list of values that you must specify before you can deploy certain applications. Some applications might include resources that can affect permissions in your AWS account, for example, by creating new AWS Identity and Access Management (IAM) users. For those applications, you must explicitly acknowledge their capabilities by specifying this parameter.

The only valid values are CAPABILITY_IAM, CAPABILITY_NAMED_IAM, CAPABILITY_RESOURCE_POLICY, and CAPABILITY_AUTO_EXPAND.

The following resources require you to specify CAPABILITY_IAM or CAPABILITY_NAMED_IAM: AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, and AWS::IAM::Role. If the application contains IAM resources, you can specify either CAPABILITY_IAM or CAPABILITY_NAMED_IAM. If the application contains IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

The following resources require you to specify CAPABILITY_RESOURCE_POLICY: AWS::Lambda::Permission, AWS::IAM:Policy, AWS::ApplicationAutoScaling::ScalingPolicy, AWS::S3::BucketPolicy, AWS::SQS::QueuePolicy, and AWS::SNS:TopicPolicy.

Applications that contain one or more nested applications require you to specify CAPABILITY_AUTO_EXPAND.

If your application template contains any of the above resources, we recommend that you review all permissions associated with the application before deploying. If you don't specify this parameter for an application that requires capabilities, the call will fail.

createCloudFormationChangeSet_parameterOverrides :: Lens' CreateCloudFormationChangeSet (Maybe [ParameterValue])

A list of parameter values for the parameters of the application.

createCloudFormationChangeSet_notificationArns :: Lens' CreateCloudFormationChangeSet (Maybe [Text])

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

createCloudFormationChangeSet_tags :: Lens' CreateCloudFormationChangeSet (Maybe [Tag])

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

createCloudFormationChangeSet_rollbackConfiguration :: Lens' CreateCloudFormationChangeSet (Maybe RollbackConfiguration)

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

createCloudFormationChangeSet_description :: Lens' CreateCloudFormationChangeSet (Maybe Text)

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

createCloudFormationChangeSet_changeSetName :: Lens' CreateCloudFormationChangeSet (Maybe Text)

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

createCloudFormationChangeSet_semanticVersion :: Lens' CreateCloudFormationChangeSet (Maybe Text)

The semantic version of the application:

https://semver.org/

createCloudFormationChangeSet_templateId :: Lens' CreateCloudFormationChangeSet (Maybe Text)

The UUID returned by CreateCloudFormationTemplate.

Pattern: [0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}

createCloudFormationChangeSet_clientToken :: Lens' CreateCloudFormationChangeSet (Maybe Text)

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

createCloudFormationChangeSet_applicationId :: Lens' CreateCloudFormationChangeSet Text

The Amazon Resource Name (ARN) of the application.

createCloudFormationChangeSet_stackName :: Lens' CreateCloudFormationChangeSet Text

This property corresponds to the parameter of the same name for the /AWS CloudFormation CreateChangeSet/ API.

createCloudFormationChangeSetResponse_applicationId :: Lens' CreateCloudFormationChangeSetResponse (Maybe Text)

The application Amazon Resource Name (ARN).

createCloudFormationChangeSetResponse_stackId :: Lens' CreateCloudFormationChangeSetResponse (Maybe Text)

The unique ID of the stack.

createCloudFormationChangeSetResponse_changeSetId :: Lens' CreateCloudFormationChangeSetResponse (Maybe Text)

The Amazon Resource Name (ARN) of the change set.

Length constraints: Minimum length of 1.

Pattern: ARN:[-a-zA-Z0-9:/]*

createCloudFormationChangeSetResponse_semanticVersion :: Lens' CreateCloudFormationChangeSetResponse (Maybe Text)

The semantic version of the application:

https://semver.org/

createCloudFormationChangeSetResponse_httpStatus :: Lens' CreateCloudFormationChangeSetResponse Int

The response's http status code.

GetApplication

getApplication_semanticVersion :: Lens' GetApplication (Maybe Text)

The semantic version of the application to get.

getApplication_applicationId :: Lens' GetApplication Text

The Amazon Resource Name (ARN) of the application.

getApplicationResponse_applicationId :: Lens' GetApplicationResponse (Maybe Text)

The application Amazon Resource Name (ARN).

getApplicationResponse_creationTime :: Lens' GetApplicationResponse (Maybe Text)

The date and time this resource was created.

getApplicationResponse_spdxLicenseId :: Lens' GetApplicationResponse (Maybe Text)

A valid identifier from https://spdx.org/licenses/.

getApplicationResponse_licenseUrl :: Lens' GetApplicationResponse (Maybe Text)

A link to a license file of the app that matches the spdxLicenseID value of your application.

Maximum size 5 MB

getApplicationResponse_verifiedAuthorUrl :: Lens' GetApplicationResponse (Maybe Text)

The URL to the public profile of a verified author. This URL is submitted by the author.

getApplicationResponse_labels :: Lens' GetApplicationResponse (Maybe [Text])

Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

getApplicationResponse_author :: Lens' GetApplicationResponse (Maybe Text)

The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

getApplicationResponse_version :: Lens' GetApplicationResponse (Maybe Version)

Version information about the application.

getApplicationResponse_name :: Lens' GetApplicationResponse (Maybe Text)

The name of the application.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

getApplicationResponse_homePageUrl :: Lens' GetApplicationResponse (Maybe Text)

A URL with more information about the application, for example the location of your GitHub repository for the application.

getApplicationResponse_readmeUrl :: Lens' GetApplicationResponse (Maybe Text)

A link to the readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

getApplicationResponse_isVerifiedAuthor :: Lens' GetApplicationResponse (Maybe Bool)

Whether the author of this application has been verified. This means means that AWS has made a good faith review, as a reasonable and prudent service provider, of the information provided by the requester and has confirmed that the requester's identity is as claimed.

getApplicationResponse_description :: Lens' GetApplicationResponse (Maybe Text)

The description of the application.

Minimum length=1. Maximum length=256

getApplicationResponse_httpStatus :: Lens' GetApplicationResponse Int

The response's http status code.

ListApplicationDependencies

listApplicationDependencies_nextToken :: Lens' ListApplicationDependencies (Maybe Text)

A token to specify where to start paginating.

listApplicationDependencies_semanticVersion :: Lens' ListApplicationDependencies (Maybe Text)

The semantic version of the application to get.

listApplicationDependencies_maxItems :: Lens' ListApplicationDependencies (Maybe Natural)

The total number of items to return.

listApplicationDependencies_applicationId :: Lens' ListApplicationDependencies Text

The Amazon Resource Name (ARN) of the application.

listApplicationDependenciesResponse_nextToken :: Lens' ListApplicationDependenciesResponse (Maybe Text)

The token to request the next page of results.

listApplicationDependenciesResponse_dependencies :: Lens' ListApplicationDependenciesResponse (Maybe [ApplicationDependencySummary])

An array of application summaries nested in the application.

listApplicationDependenciesResponse_httpStatus :: Lens' ListApplicationDependenciesResponse Int

The response's http status code.

ListApplicationVersions

listApplicationVersions_nextToken :: Lens' ListApplicationVersions (Maybe Text)

A token to specify where to start paginating.

listApplicationVersions_maxItems :: Lens' ListApplicationVersions (Maybe Natural)

The total number of items to return.

listApplicationVersions_applicationId :: Lens' ListApplicationVersions Text

The Amazon Resource Name (ARN) of the application.

listApplicationVersionsResponse_nextToken :: Lens' ListApplicationVersionsResponse (Maybe Text)

The token to request the next page of results.

listApplicationVersionsResponse_versions :: Lens' ListApplicationVersionsResponse (Maybe [VersionSummary])

An array of version summaries for the application.

listApplicationVersionsResponse_httpStatus :: Lens' ListApplicationVersionsResponse Int

The response's http status code.

CreateCloudFormationTemplate

createCloudFormationTemplate_semanticVersion :: Lens' CreateCloudFormationTemplate (Maybe Text)

The semantic version of the application:

https://semver.org/

createCloudFormationTemplate_applicationId :: Lens' CreateCloudFormationTemplate Text

The Amazon Resource Name (ARN) of the application.

createCloudFormationTemplateResponse_applicationId :: Lens' CreateCloudFormationTemplateResponse (Maybe Text)

The application Amazon Resource Name (ARN).

createCloudFormationTemplateResponse_creationTime :: Lens' CreateCloudFormationTemplateResponse (Maybe Text)

The date and time this resource was created.

createCloudFormationTemplateResponse_status :: Lens' CreateCloudFormationTemplateResponse (Maybe Status)

Status of the template creation workflow.

Possible values: PREPARING | ACTIVE | EXPIRED

createCloudFormationTemplateResponse_expirationTime :: Lens' CreateCloudFormationTemplateResponse (Maybe Text)

The date and time this template expires. Templates expire 1 hour after creation.

createCloudFormationTemplateResponse_templateUrl :: Lens' CreateCloudFormationTemplateResponse (Maybe Text)

A link to the template that can be used to deploy the application using AWS CloudFormation.

createCloudFormationTemplateResponse_semanticVersion :: Lens' CreateCloudFormationTemplateResponse (Maybe Text)

The semantic version of the application:

https://semver.org/

createCloudFormationTemplateResponse_templateId :: Lens' CreateCloudFormationTemplateResponse (Maybe Text)

The UUID returned by CreateCloudFormationTemplate.

Pattern: [0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}

createCloudFormationTemplateResponse_httpStatus :: Lens' CreateCloudFormationTemplateResponse Int

The response's http status code.

ListApplications

listApplications_nextToken :: Lens' ListApplications (Maybe Text)

A token to specify where to start paginating.

listApplications_maxItems :: Lens' ListApplications (Maybe Natural)

The total number of items to return.

listApplicationsResponse_nextToken :: Lens' ListApplicationsResponse (Maybe Text)

The token to request the next page of results.

listApplicationsResponse_applications :: Lens' ListApplicationsResponse (Maybe [ApplicationSummary])

An array of application summaries.

listApplicationsResponse_httpStatus :: Lens' ListApplicationsResponse Int

The response's http status code.

DeleteApplication

deleteApplication_applicationId :: Lens' DeleteApplication Text

The Amazon Resource Name (ARN) of the application.

UpdateApplication

updateApplication_labels :: Lens' UpdateApplication (Maybe [Text])

Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

updateApplication_author :: Lens' UpdateApplication (Maybe Text)

The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

updateApplication_readmeBody :: Lens' UpdateApplication (Maybe Text)

A text readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

updateApplication_homePageUrl :: Lens' UpdateApplication (Maybe Text)

A URL with more information about the application, for example the location of your GitHub repository for the application.

updateApplication_readmeUrl :: Lens' UpdateApplication (Maybe Text)

A link to the readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

updateApplication_description :: Lens' UpdateApplication (Maybe Text)

The description of the application.

Minimum length=1. Maximum length=256

updateApplication_applicationId :: Lens' UpdateApplication Text

The Amazon Resource Name (ARN) of the application.

updateApplicationResponse_applicationId :: Lens' UpdateApplicationResponse (Maybe Text)

The application Amazon Resource Name (ARN).

updateApplicationResponse_creationTime :: Lens' UpdateApplicationResponse (Maybe Text)

The date and time this resource was created.

updateApplicationResponse_spdxLicenseId :: Lens' UpdateApplicationResponse (Maybe Text)

A valid identifier from https://spdx.org/licenses/.

updateApplicationResponse_licenseUrl :: Lens' UpdateApplicationResponse (Maybe Text)

A link to a license file of the app that matches the spdxLicenseID value of your application.

Maximum size 5 MB

updateApplicationResponse_verifiedAuthorUrl :: Lens' UpdateApplicationResponse (Maybe Text)

The URL to the public profile of a verified author. This URL is submitted by the author.

updateApplicationResponse_labels :: Lens' UpdateApplicationResponse (Maybe [Text])

Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

updateApplicationResponse_author :: Lens' UpdateApplicationResponse (Maybe Text)

The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

updateApplicationResponse_version :: Lens' UpdateApplicationResponse (Maybe Version)

Version information about the application.

updateApplicationResponse_name :: Lens' UpdateApplicationResponse (Maybe Text)

The name of the application.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

updateApplicationResponse_homePageUrl :: Lens' UpdateApplicationResponse (Maybe Text)

A URL with more information about the application, for example the location of your GitHub repository for the application.

updateApplicationResponse_readmeUrl :: Lens' UpdateApplicationResponse (Maybe Text)

A link to the readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

updateApplicationResponse_isVerifiedAuthor :: Lens' UpdateApplicationResponse (Maybe Bool)

Whether the author of this application has been verified. This means means that AWS has made a good faith review, as a reasonable and prudent service provider, of the information provided by the requester and has confirmed that the requester's identity is as claimed.

updateApplicationResponse_description :: Lens' UpdateApplicationResponse (Maybe Text)

The description of the application.

Minimum length=1. Maximum length=256

updateApplicationResponse_httpStatus :: Lens' UpdateApplicationResponse Int

The response's http status code.

PutApplicationPolicy

putApplicationPolicy_applicationId :: Lens' PutApplicationPolicy Text

The Amazon Resource Name (ARN) of the application.

putApplicationPolicy_statements :: Lens' PutApplicationPolicy [ApplicationPolicyStatement]

An array of policy statements applied to the application.

putApplicationPolicyResponse_statements :: Lens' PutApplicationPolicyResponse (Maybe [ApplicationPolicyStatement])

An array of policy statements applied to the application.

putApplicationPolicyResponse_httpStatus :: Lens' PutApplicationPolicyResponse Int

The response's http status code.

Types

ApplicationDependencySummary

applicationDependencySummary_applicationId :: Lens' ApplicationDependencySummary Text

The Amazon Resource Name (ARN) of the nested application.

applicationDependencySummary_semanticVersion :: Lens' ApplicationDependencySummary Text

The semantic version of the nested application.

ApplicationPolicyStatement

applicationPolicyStatement_statementId :: Lens' ApplicationPolicyStatement (Maybe Text)

A unique ID for the statement.

applicationPolicyStatement_principalOrgIDs :: Lens' ApplicationPolicyStatement (Maybe [Text])

An array of PrinciplalOrgIDs, which corresponds to AWS IAM aws:PrincipalOrgID global condition key.

applicationPolicyStatement_principals :: Lens' ApplicationPolicyStatement [Text]

An array of AWS account IDs, or * to make the application public.

applicationPolicyStatement_actions :: Lens' ApplicationPolicyStatement [Text]

For the list of actions supported for this operation, see Application Permissions.

ApplicationSummary

applicationSummary_creationTime :: Lens' ApplicationSummary (Maybe Text)

The date and time this resource was created.

applicationSummary_spdxLicenseId :: Lens' ApplicationSummary (Maybe Text)

A valid identifier from https://spdx.org/licenses/.

applicationSummary_labels :: Lens' ApplicationSummary (Maybe [Text])

Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

applicationSummary_homePageUrl :: Lens' ApplicationSummary (Maybe Text)

A URL with more information about the application, for example the location of your GitHub repository for the application.

applicationSummary_description :: Lens' ApplicationSummary Text

The description of the application.

Minimum length=1. Maximum length=256

applicationSummary_author :: Lens' ApplicationSummary Text

The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

applicationSummary_applicationId :: Lens' ApplicationSummary Text

The application Amazon Resource Name (ARN).

applicationSummary_name :: Lens' ApplicationSummary Text

The name of the application.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

ParameterDefinition

parameterDefinition_maxValue :: Lens' ParameterDefinition (Maybe Int)

A numeric value that determines the largest numeric value that you want to allow for Number types.

parameterDefinition_minLength :: Lens' ParameterDefinition (Maybe Int)

An integer value that determines the smallest number of characters that you want to allow for String types.

parameterDefinition_allowedValues :: Lens' ParameterDefinition (Maybe [Text])

An array containing the list of values allowed for the parameter.

parameterDefinition_minValue :: Lens' ParameterDefinition (Maybe Int)

A numeric value that determines the smallest numeric value that you want to allow for Number types.

parameterDefinition_description :: Lens' ParameterDefinition (Maybe Text)

A string of up to 4,000 characters that describes the parameter.

parameterDefinition_constraintDescription :: Lens' ParameterDefinition (Maybe Text)

A string that explains a constraint when the constraint is violated. For example, without a constraint description, a parameter that has an allowed pattern of [A-Za-z0-9]+ displays the following error message when the user specifies an invalid value:

Malformed input-Parameter MyParameter must match pattern [A-Za-z0-9]+

By adding a constraint description, such as "must contain only uppercase and lowercase letters and numbers," you can display the following customized error message:

Malformed input-Parameter MyParameter must contain only uppercase and lowercase letters and numbers.

parameterDefinition_maxLength :: Lens' ParameterDefinition (Maybe Int)

An integer value that determines the largest number of characters that you want to allow for String types.

parameterDefinition_defaultValue :: Lens' ParameterDefinition (Maybe Text)

A value of the appropriate type for the template to use if no value is specified when a stack is created. If you define constraints for the parameter, you must specify a value that adheres to those constraints.

parameterDefinition_allowedPattern :: Lens' ParameterDefinition (Maybe Text)

A regular expression that represents the patterns to allow for String types.

parameterDefinition_type :: Lens' ParameterDefinition (Maybe Text)

The type of the parameter.

Valid values: String | Number | List<Number> | CommaDelimitedList

String: A literal string.

For example, users can specify "MyUserName".

Number: An integer or float. AWS CloudFormation validates the parameter value as a number. However, when you use the parameter elsewhere in your template (for example, by using the Ref intrinsic function), the parameter value becomes a string.

For example, users might specify "8888".

List<Number>: An array of integers or floats that are separated by commas. AWS CloudFormation validates the parameter value as numbers. However, when you use the parameter elsewhere in your template (for example, by using the Ref intrinsic function), the parameter value becomes a list of strings.

For example, users might specify "80,20", and then Ref results in ["80","20"].

CommaDelimitedList: An array of literal strings that are separated by commas. The total number of strings should be one more than the total number of commas. Also, each member string is space-trimmed.

For example, users might specify "test,dev,prod", and then Ref results in ["test","dev","prod"].

parameterDefinition_noEcho :: Lens' ParameterDefinition (Maybe Bool)

Whether to mask the parameter value whenever anyone makes a call that describes the stack. If you set the value to true, the parameter value is masked with asterisks (*****).

parameterDefinition_referencedByResources :: Lens' ParameterDefinition [Text]

A list of AWS SAM resources that use this parameter.

parameterDefinition_name :: Lens' ParameterDefinition Text

The name of the parameter.

ParameterValue

parameterValue_value :: Lens' ParameterValue Text

The input value associated with the parameter.

parameterValue_name :: Lens' ParameterValue Text

The key associated with the parameter. If you don't specify a key and value for a particular parameter, AWS CloudFormation uses the default value that is specified in your template.

RollbackConfiguration

rollbackConfiguration_monitoringTimeInMinutes :: Lens' RollbackConfiguration (Maybe Int)

This property corresponds to the content of the same name for the /AWS CloudFormation RollbackConfiguration/ Data Type.

rollbackConfiguration_rollbackTriggers :: Lens' RollbackConfiguration (Maybe [RollbackTrigger])

This property corresponds to the content of the same name for the /AWS CloudFormation RollbackConfiguration/ Data Type.

RollbackTrigger

rollbackTrigger_type :: Lens' RollbackTrigger Text

This property corresponds to the content of the same name for the /AWS CloudFormation RollbackTrigger/ Data Type.

rollbackTrigger_arn :: Lens' RollbackTrigger Text

This property corresponds to the content of the same name for the /AWS CloudFormation RollbackTrigger/ Data Type.

Tag

tag_value :: Lens' Tag Text

This property corresponds to the content of the same name for the /AWS CloudFormation Tag/ Data Type.

tag_key :: Lens' Tag Text

This property corresponds to the content of the same name for the /AWS CloudFormation Tag/ Data Type.

Version

version_sourceCodeArchiveUrl :: Lens' Version (Maybe Text)

A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

version_sourceCodeUrl :: Lens' Version (Maybe Text)

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

version_templateUrl :: Lens' Version Text

A link to the packaged AWS SAM template of your application.

version_parameterDefinitions :: Lens' Version [ParameterDefinition]

An array of parameter types supported by the application.

version_resourcesSupported :: Lens' Version Bool

Whether all of the AWS resources contained in this application are supported in the region in which it is being retrieved.

version_creationTime :: Lens' Version Text

The date and time this resource was created.

version_requiredCapabilities :: Lens' Version [Capability]

A list of values that you must specify before you can deploy certain applications. Some applications might include resources that can affect permissions in your AWS account, for example, by creating new AWS Identity and Access Management (IAM) users. For those applications, you must explicitly acknowledge their capabilities by specifying this parameter.

The only valid values are CAPABILITY_IAM, CAPABILITY_NAMED_IAM, CAPABILITY_RESOURCE_POLICY, and CAPABILITY_AUTO_EXPAND.

The following resources require you to specify CAPABILITY_IAM or CAPABILITY_NAMED_IAM: AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, and AWS::IAM::Role. If the application contains IAM resources, you can specify either CAPABILITY_IAM or CAPABILITY_NAMED_IAM. If the application contains IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

The following resources require you to specify CAPABILITY_RESOURCE_POLICY: AWS::Lambda::Permission, AWS::IAM:Policy, AWS::ApplicationAutoScaling::ScalingPolicy, AWS::S3::BucketPolicy, AWS::SQS::QueuePolicy, and AWS::SNS::TopicPolicy.

Applications that contain one or more nested applications require you to specify CAPABILITY_AUTO_EXPAND.

If your application template contains any of the above resources, we recommend that you review all permissions associated with the application before deploying. If you don't specify this parameter for an application that requires capabilities, the call will fail.

version_applicationId :: Lens' Version Text

The application Amazon Resource Name (ARN).

version_semanticVersion :: Lens' Version Text

The semantic version of the application:

https://semver.org/

VersionSummary

versionSummary_sourceCodeUrl :: Lens' VersionSummary (Maybe Text)

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

versionSummary_creationTime :: Lens' VersionSummary Text

The date and time this resource was created.

versionSummary_applicationId :: Lens' VersionSummary Text

The application Amazon Resource Name (ARN).

versionSummary_semanticVersion :: Lens' VersionSummary Text

The semantic version of the application:

https://semver.org/