Network.AWS.ServerlessApplicationRepository.Types

Description

Service Configuration

defaultService :: Service

API version 2017-09-08 of the Amazon ServerlessApplicationRepository SDK configuration.

Errors

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resource (for example, an access policy statement) specified in the request doesn't exist.

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

One of the parameters in the request is invalid.

_InternalServerErrorException :: AsError a => Getting (First ServiceError) a ServiceError

The AWS Serverless Application Repository service encountered an internal error.

_ForbiddenException :: AsError a => Getting (First ServiceError) a ServiceError

The client is not authenticated.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The resource already exists.

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

The client is sending more than the allowed number of requests per unit of time.

Capability

newtype Capability

Values that must be specified in order to deploy some applications.

Constructors

Capability'

Fields

fromCapability :: Text

Bundled Patterns

pattern Capability_CAPABILITY_AUTO_EXPAND :: Capability

pattern Capability_CAPABILITY_IAM :: Capability

pattern Capability_CAPABILITY_NAMED_IAM :: Capability

pattern Capability_CAPABILITY_RESOURCE_POLICY :: Capability

Status

newtype Status

Constructors

Status'

Fields

fromStatus :: Text

Bundled Patterns

pattern Status_ACTIVE :: Status

pattern Status_EXPIRED :: Status

pattern Status_PREPARING :: Status

ApplicationDependencySummary

data ApplicationDependencySummary

A nested application summary.

See: newApplicationDependencySummary smart constructor.

Constructors

ApplicationDependencySummary'

Fields

applicationId :: Text

The Amazon Resource Name (ARN) of the nested application.

semanticVersion :: Text

The semantic version of the nested application.

newApplicationDependencySummary

Arguments

:: Text
$sel:applicationId:ApplicationDependencySummary'

-> Text
$sel:semanticVersion:ApplicationDependencySummary'

-> ApplicationDependencySummary

Create a value of ApplicationDependencySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:ApplicationDependencySummary', applicationDependencySummary_applicationId - The Amazon Resource Name (ARN) of the nested application.

$sel:semanticVersion:ApplicationDependencySummary', applicationDependencySummary_semanticVersion - The semantic version of the nested application.

applicationDependencySummary_applicationId :: Lens' ApplicationDependencySummary Text

The Amazon Resource Name (ARN) of the nested application.

applicationDependencySummary_semanticVersion :: Lens' ApplicationDependencySummary Text

The semantic version of the nested application.

ApplicationPolicyStatement

data ApplicationPolicyStatement

Policy statement applied to the application.

See: newApplicationPolicyStatement smart constructor.

Constructors

ApplicationPolicyStatement'

Fields

statementId :: Maybe Text

A unique ID for the statement.

principalOrgIDs :: Maybe [Text]

An array of PrinciplalOrgIDs, which corresponds to AWS IAM aws:PrincipalOrgID global condition key.

principals :: [Text]

An array of AWS account IDs, or * to make the application public.

actions :: [Text]

For the list of actions supported for this operation, see Application Permissions.

newApplicationPolicyStatement :: ApplicationPolicyStatement

Create a value of ApplicationPolicyStatement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statementId:ApplicationPolicyStatement', applicationPolicyStatement_statementId - A unique ID for the statement.

$sel:principalOrgIDs:ApplicationPolicyStatement', applicationPolicyStatement_principalOrgIDs - An array of PrinciplalOrgIDs, which corresponds to AWS IAM aws:PrincipalOrgID global condition key.

$sel:principals:ApplicationPolicyStatement', applicationPolicyStatement_principals - An array of AWS account IDs, or * to make the application public.

$sel:actions:ApplicationPolicyStatement', applicationPolicyStatement_actions - For the list of actions supported for this operation, see Application Permissions.

applicationPolicyStatement_statementId :: Lens' ApplicationPolicyStatement (Maybe Text)

A unique ID for the statement.

applicationPolicyStatement_principalOrgIDs :: Lens' ApplicationPolicyStatement (Maybe [Text])

An array of PrinciplalOrgIDs, which corresponds to AWS IAM aws:PrincipalOrgID global condition key.

applicationPolicyStatement_principals :: Lens' ApplicationPolicyStatement [Text]

An array of AWS account IDs, or * to make the application public.

applicationPolicyStatement_actions :: Lens' ApplicationPolicyStatement [Text]

For the list of actions supported for this operation, see Application Permissions.

ApplicationSummary

data ApplicationSummary

Summary of details about the application.

See: newApplicationSummary smart constructor.

Constructors

ApplicationSummary'

Fields

creationTime :: Maybe Text

The date and time this resource was created.

spdxLicenseId :: Maybe Text

A valid identifier from https://spdx.org/licenses/.

labels :: Maybe [Text]

Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

homePageUrl :: Maybe Text

A URL with more information about the application, for example the location of your GitHub repository for the application.

description :: Text

The description of the application.

Minimum length=1. Maximum length=256

author :: Text

The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

applicationId :: Text

The application Amazon Resource Name (ARN).

name :: Text

The name of the application.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

newApplicationSummary

Arguments

:: Text
$sel:description:ApplicationSummary'

-> Text
$sel:author:ApplicationSummary'

-> Text
$sel:applicationId:ApplicationSummary'

-> Text
$sel:name:ApplicationSummary'

-> ApplicationSummary

Create a value of ApplicationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ApplicationSummary', applicationSummary_creationTime - The date and time this resource was created.

$sel:spdxLicenseId:ApplicationSummary', applicationSummary_spdxLicenseId - A valid identifier from https://spdx.org/licenses/.

$sel:labels:ApplicationSummary', applicationSummary_labels - Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

$sel:homePageUrl:ApplicationSummary', applicationSummary_homePageUrl - A URL with more information about the application, for example the location of your GitHub repository for the application.

$sel:description:ApplicationSummary', applicationSummary_description - The description of the application.

Minimum length=1. Maximum length=256

$sel:author:ApplicationSummary', applicationSummary_author - The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

$sel:applicationId:ApplicationSummary', applicationSummary_applicationId - The application Amazon Resource Name (ARN).

$sel:name:ApplicationSummary', applicationSummary_name - The name of the application.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

applicationSummary_creationTime :: Lens' ApplicationSummary (Maybe Text)

The date and time this resource was created.

applicationSummary_spdxLicenseId :: Lens' ApplicationSummary (Maybe Text)

A valid identifier from https://spdx.org/licenses/.

applicationSummary_labels :: Lens' ApplicationSummary (Maybe [Text])

Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

applicationSummary_homePageUrl :: Lens' ApplicationSummary (Maybe Text)

A URL with more information about the application, for example the location of your GitHub repository for the application.

applicationSummary_description :: Lens' ApplicationSummary Text

The description of the application.

Minimum length=1. Maximum length=256

applicationSummary_author :: Lens' ApplicationSummary Text

The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

applicationSummary_applicationId :: Lens' ApplicationSummary Text

The application Amazon Resource Name (ARN).

applicationSummary_name :: Lens' ApplicationSummary Text

The name of the application.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

ParameterDefinition

data ParameterDefinition

Parameters supported by the application.

See: newParameterDefinition smart constructor.

Constructors

ParameterDefinition'

Fields

maxValue :: Maybe Int

A numeric value that determines the largest numeric value that you want to allow for Number types.

minLength :: Maybe Int

An integer value that determines the smallest number of characters that you want to allow for String types.

allowedValues :: Maybe [Text]

An array containing the list of values allowed for the parameter.

minValue :: Maybe Int

A numeric value that determines the smallest numeric value that you want to allow for Number types.

description :: Maybe Text

A string of up to 4,000 characters that describes the parameter.

constraintDescription :: Maybe Text

A string that explains a constraint when the constraint is violated. For example, without a constraint description, a parameter that has an allowed pattern of [A-Za-z0-9]+ displays the following error message when the user specifies an invalid value:

Malformed input-Parameter MyParameter must match pattern [A-Za-z0-9]+

By adding a constraint description, such as "must contain only uppercase and lowercase letters and numbers," you can display the following customized error message:

Malformed input-Parameter MyParameter must contain only uppercase and lowercase letters and numbers.

maxLength :: Maybe Int

An integer value that determines the largest number of characters that you want to allow for String types.

defaultValue :: Maybe Text

A value of the appropriate type for the template to use if no value is specified when a stack is created. If you define constraints for the parameter, you must specify a value that adheres to those constraints.

allowedPattern :: Maybe Text

A regular expression that represents the patterns to allow for String types.

type' :: Maybe Text

The type of the parameter.

Valid values: String | Number | List<Number> | CommaDelimitedList

String: A literal string.

For example, users can specify "MyUserName".

Number: An integer or float. AWS CloudFormation validates the parameter value as a number. However, when you use the parameter elsewhere in your template (for example, by using the Ref intrinsic function), the parameter value becomes a string.

For example, users might specify "8888".

List<Number>: An array of integers or floats that are separated by commas. AWS CloudFormation validates the parameter value as numbers. However, when you use the parameter elsewhere in your template (for example, by using the Ref intrinsic function), the parameter value becomes a list of strings.

For example, users might specify "80,20", and then Ref results in ["80","20"].

CommaDelimitedList: An array of literal strings that are separated by commas. The total number of strings should be one more than the total number of commas. Also, each member string is space-trimmed.

For example, users might specify "test,dev,prod", and then Ref results in ["test","dev","prod"].

noEcho :: Maybe Bool

Whether to mask the parameter value whenever anyone makes a call that describes the stack. If you set the value to true, the parameter value is masked with asterisks (*****).

referencedByResources :: [Text]

A list of AWS SAM resources that use this parameter.

name :: Text

The name of the parameter.

newParameterDefinition

Arguments

:: Text
$sel:name:ParameterDefinition'

-> ParameterDefinition

Create a value of ParameterDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxValue:ParameterDefinition', parameterDefinition_maxValue - A numeric value that determines the largest numeric value that you want to allow for Number types.

$sel:minLength:ParameterDefinition', parameterDefinition_minLength - An integer value that determines the smallest number of characters that you want to allow for String types.

$sel:allowedValues:ParameterDefinition', parameterDefinition_allowedValues - An array containing the list of values allowed for the parameter.

$sel:minValue:ParameterDefinition', parameterDefinition_minValue - A numeric value that determines the smallest numeric value that you want to allow for Number types.

$sel:description:ParameterDefinition', parameterDefinition_description - A string of up to 4,000 characters that describes the parameter.

$sel:constraintDescription:ParameterDefinition', parameterDefinition_constraintDescription - A string that explains a constraint when the constraint is violated. For example, without a constraint description, a parameter that has an allowed pattern of [A-Za-z0-9]+ displays the following error message when the user specifies an invalid value:

Malformed input-Parameter MyParameter must match pattern [A-Za-z0-9]+

By adding a constraint description, such as "must contain only uppercase and lowercase letters and numbers," you can display the following customized error message:

Malformed input-Parameter MyParameter must contain only uppercase and lowercase letters and numbers.

$sel:maxLength:ParameterDefinition', parameterDefinition_maxLength - An integer value that determines the largest number of characters that you want to allow for String types.

$sel:defaultValue:ParameterDefinition', parameterDefinition_defaultValue - A value of the appropriate type for the template to use if no value is specified when a stack is created. If you define constraints for the parameter, you must specify a value that adheres to those constraints.

$sel:allowedPattern:ParameterDefinition', parameterDefinition_allowedPattern - A regular expression that represents the patterns to allow for String types.

$sel:type':ParameterDefinition', parameterDefinition_type - The type of the parameter.

Valid values: String | Number | List<Number> | CommaDelimitedList

String: A literal string.

For example, users can specify "MyUserName".

Number: An integer or float. AWS CloudFormation validates the parameter value as a number. However, when you use the parameter elsewhere in your template (for example, by using the Ref intrinsic function), the parameter value becomes a string.

For example, users might specify "8888".

List<Number>: An array of integers or floats that are separated by commas. AWS CloudFormation validates the parameter value as numbers. However, when you use the parameter elsewhere in your template (for example, by using the Ref intrinsic function), the parameter value becomes a list of strings.

For example, users might specify "80,20", and then Ref results in ["80","20"].

CommaDelimitedList: An array of literal strings that are separated by commas. The total number of strings should be one more than the total number of commas. Also, each member string is space-trimmed.

For example, users might specify "test,dev,prod", and then Ref results in ["test","dev","prod"].

$sel:noEcho:ParameterDefinition', parameterDefinition_noEcho - Whether to mask the parameter value whenever anyone makes a call that describes the stack. If you set the value to true, the parameter value is masked with asterisks (*****).

$sel:referencedByResources:ParameterDefinition', parameterDefinition_referencedByResources - A list of AWS SAM resources that use this parameter.

$sel:name:ParameterDefinition', parameterDefinition_name - The name of the parameter.

parameterDefinition_maxValue :: Lens' ParameterDefinition (Maybe Int)

A numeric value that determines the largest numeric value that you want to allow for Number types.

parameterDefinition_minLength :: Lens' ParameterDefinition (Maybe Int)

An integer value that determines the smallest number of characters that you want to allow for String types.

parameterDefinition_allowedValues :: Lens' ParameterDefinition (Maybe [Text])

An array containing the list of values allowed for the parameter.

parameterDefinition_minValue :: Lens' ParameterDefinition (Maybe Int)

A numeric value that determines the smallest numeric value that you want to allow for Number types.

parameterDefinition_description :: Lens' ParameterDefinition (Maybe Text)

A string of up to 4,000 characters that describes the parameter.

parameterDefinition_constraintDescription :: Lens' ParameterDefinition (Maybe Text)

A string that explains a constraint when the constraint is violated. For example, without a constraint description, a parameter that has an allowed pattern of [A-Za-z0-9]+ displays the following error message when the user specifies an invalid value:

Malformed input-Parameter MyParameter must match pattern [A-Za-z0-9]+

By adding a constraint description, such as "must contain only uppercase and lowercase letters and numbers," you can display the following customized error message:

Malformed input-Parameter MyParameter must contain only uppercase and lowercase letters and numbers.

parameterDefinition_maxLength :: Lens' ParameterDefinition (Maybe Int)

An integer value that determines the largest number of characters that you want to allow for String types.

parameterDefinition_defaultValue :: Lens' ParameterDefinition (Maybe Text)

A value of the appropriate type for the template to use if no value is specified when a stack is created. If you define constraints for the parameter, you must specify a value that adheres to those constraints.

parameterDefinition_allowedPattern :: Lens' ParameterDefinition (Maybe Text)

A regular expression that represents the patterns to allow for String types.

parameterDefinition_type :: Lens' ParameterDefinition (Maybe Text)

The type of the parameter.

Valid values: String | Number | List<Number> | CommaDelimitedList

String: A literal string.

For example, users can specify "MyUserName".

Number: An integer or float. AWS CloudFormation validates the parameter value as a number. However, when you use the parameter elsewhere in your template (for example, by using the Ref intrinsic function), the parameter value becomes a string.

For example, users might specify "8888".

List<Number>: An array of integers or floats that are separated by commas. AWS CloudFormation validates the parameter value as numbers. However, when you use the parameter elsewhere in your template (for example, by using the Ref intrinsic function), the parameter value becomes a list of strings.

For example, users might specify "80,20", and then Ref results in ["80","20"].

CommaDelimitedList: An array of literal strings that are separated by commas. The total number of strings should be one more than the total number of commas. Also, each member string is space-trimmed.

For example, users might specify "test,dev,prod", and then Ref results in ["test","dev","prod"].

parameterDefinition_noEcho :: Lens' ParameterDefinition (Maybe Bool)

Whether to mask the parameter value whenever anyone makes a call that describes the stack. If you set the value to true, the parameter value is masked with asterisks (*****).

parameterDefinition_referencedByResources :: Lens' ParameterDefinition [Text]

A list of AWS SAM resources that use this parameter.

parameterDefinition_name :: Lens' ParameterDefinition Text

The name of the parameter.

ParameterValue

data ParameterValue

Parameter value of the application.

See: newParameterValue smart constructor.

Constructors

ParameterValue'

Fields

value :: Text

The input value associated with the parameter.

name :: Text

The key associated with the parameter. If you don't specify a key and value for a particular parameter, AWS CloudFormation uses the default value that is specified in your template.

newParameterValue

Arguments

:: Text
$sel:value:ParameterValue'

-> Text
$sel:name:ParameterValue'

-> ParameterValue

Create a value of ParameterValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:ParameterValue', parameterValue_value - The input value associated with the parameter.

$sel:name:ParameterValue', parameterValue_name - The key associated with the parameter. If you don't specify a key and value for a particular parameter, AWS CloudFormation uses the default value that is specified in your template.

parameterValue_value :: Lens' ParameterValue Text

The input value associated with the parameter.

parameterValue_name :: Lens' ParameterValue Text

The key associated with the parameter. If you don't specify a key and value for a particular parameter, AWS CloudFormation uses the default value that is specified in your template.

RollbackConfiguration

data RollbackConfiguration

This property corresponds to the /AWS CloudFormation RollbackConfiguration/ Data Type.

See: newRollbackConfiguration smart constructor.

Constructors

RollbackConfiguration'

Fields

monitoringTimeInMinutes :: Maybe Int

This property corresponds to the content of the same name for the /AWS CloudFormation RollbackConfiguration/ Data Type.

rollbackTriggers :: Maybe [RollbackTrigger]

This property corresponds to the content of the same name for the /AWS CloudFormation RollbackConfiguration/ Data Type.

newRollbackConfiguration :: RollbackConfiguration

Create a value of RollbackConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:monitoringTimeInMinutes:RollbackConfiguration', rollbackConfiguration_monitoringTimeInMinutes - This property corresponds to the content of the same name for the /AWS CloudFormation RollbackConfiguration/ Data Type.

$sel:rollbackTriggers:RollbackConfiguration', rollbackConfiguration_rollbackTriggers - This property corresponds to the content of the same name for the /AWS CloudFormation RollbackConfiguration/ Data Type.

rollbackConfiguration_monitoringTimeInMinutes :: Lens' RollbackConfiguration (Maybe Int)

This property corresponds to the content of the same name for the /AWS CloudFormation RollbackConfiguration/ Data Type.

rollbackConfiguration_rollbackTriggers :: Lens' RollbackConfiguration (Maybe [RollbackTrigger])

This property corresponds to the content of the same name for the /AWS CloudFormation RollbackConfiguration/ Data Type.

RollbackTrigger

data RollbackTrigger

This property corresponds to the /AWS CloudFormation RollbackTrigger/ Data Type.

See: newRollbackTrigger smart constructor.

Constructors

RollbackTrigger'

Fields

type' :: Text

This property corresponds to the content of the same name for the /AWS CloudFormation RollbackTrigger/ Data Type.

arn :: Text

This property corresponds to the content of the same name for the /AWS CloudFormation RollbackTrigger/ Data Type.

newRollbackTrigger

Arguments

:: Text
$sel:type':RollbackTrigger'

-> Text
$sel:arn:RollbackTrigger'

-> RollbackTrigger

Create a value of RollbackTrigger with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':RollbackTrigger', rollbackTrigger_type - This property corresponds to the content of the same name for the /AWS CloudFormation RollbackTrigger/ Data Type.

$sel:arn:RollbackTrigger', rollbackTrigger_arn - This property corresponds to the content of the same name for the /AWS CloudFormation RollbackTrigger/ Data Type.

rollbackTrigger_type :: Lens' RollbackTrigger Text

This property corresponds to the content of the same name for the /AWS CloudFormation RollbackTrigger/ Data Type.

rollbackTrigger_arn :: Lens' RollbackTrigger Text

This property corresponds to the content of the same name for the /AWS CloudFormation RollbackTrigger/ Data Type.

Tag

data Tag

This property corresponds to the /AWS CloudFormation Tag/ Data Type.

See: newTag smart constructor.

Constructors

Tag'

Fields

value :: Text

This property corresponds to the content of the same name for the /AWS CloudFormation Tag/ Data Type.

key :: Text

This property corresponds to the content of the same name for the /AWS CloudFormation Tag/ Data Type.

newTag

Arguments

:: Text
$sel:value:Tag'

-> Text
$sel:key:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:Tag', tag_value - This property corresponds to the content of the same name for the /AWS CloudFormation Tag/ Data Type.

$sel:key:Tag', tag_key - This property corresponds to the content of the same name for the /AWS CloudFormation Tag/ Data Type.

tag_value :: Lens' Tag Text

This property corresponds to the content of the same name for the /AWS CloudFormation Tag/ Data Type.

tag_key :: Lens' Tag Text

This property corresponds to the content of the same name for the /AWS CloudFormation Tag/ Data Type.

Version

data Version

Application version details.

See: newVersion smart constructor.

Constructors

Version'

Fields

sourceCodeArchiveUrl :: Maybe Text

A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

sourceCodeUrl :: Maybe Text

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

templateUrl :: Text

A link to the packaged AWS SAM template of your application.

parameterDefinitions :: [ParameterDefinition]

An array of parameter types supported by the application.

resourcesSupported :: Bool

Whether all of the AWS resources contained in this application are supported in the region in which it is being retrieved.

creationTime :: Text

The date and time this resource was created.

requiredCapabilities :: [Capability]

A list of values that you must specify before you can deploy certain applications. Some applications might include resources that can affect permissions in your AWS account, for example, by creating new AWS Identity and Access Management (IAM) users. For those applications, you must explicitly acknowledge their capabilities by specifying this parameter.

The only valid values are CAPABILITY_IAM, CAPABILITY_NAMED_IAM, CAPABILITY_RESOURCE_POLICY, and CAPABILITY_AUTO_EXPAND.

The following resources require you to specify CAPABILITY_IAM or CAPABILITY_NAMED_IAM: AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, and AWS::IAM::Role. If the application contains IAM resources, you can specify either CAPABILITY_IAM or CAPABILITY_NAMED_IAM. If the application contains IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

The following resources require you to specify CAPABILITY_RESOURCE_POLICY: AWS::Lambda::Permission, AWS::IAM:Policy, AWS::ApplicationAutoScaling::ScalingPolicy, AWS::S3::BucketPolicy, AWS::SQS::QueuePolicy, and AWS::SNS::TopicPolicy.

Applications that contain one or more nested applications require you to specify CAPABILITY_AUTO_EXPAND.

If your application template contains any of the above resources, we recommend that you review all permissions associated with the application before deploying. If you don't specify this parameter for an application that requires capabilities, the call will fail.

applicationId :: Text

The application Amazon Resource Name (ARN).

semanticVersion :: Text

The semantic version of the application:

https://semver.org/

newVersion

Arguments

:: Text
$sel:templateUrl:Version'

-> Bool
$sel:resourcesSupported:Version'

-> Text
$sel:creationTime:Version'

-> Text
$sel:applicationId:Version'

-> Text
$sel:semanticVersion:Version'

-> Version

Create a value of Version with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceCodeArchiveUrl:Version', version_sourceCodeArchiveUrl - A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

$sel:sourceCodeUrl:Version', version_sourceCodeUrl - A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

$sel:templateUrl:Version', version_templateUrl - A link to the packaged AWS SAM template of your application.

$sel:parameterDefinitions:Version', version_parameterDefinitions - An array of parameter types supported by the application.

$sel:resourcesSupported:Version', version_resourcesSupported - Whether all of the AWS resources contained in this application are supported in the region in which it is being retrieved.

$sel:creationTime:Version', version_creationTime - The date and time this resource was created.

$sel:requiredCapabilities:Version', version_requiredCapabilities - A list of values that you must specify before you can deploy certain applications. Some applications might include resources that can affect permissions in your AWS account, for example, by creating new AWS Identity and Access Management (IAM) users. For those applications, you must explicitly acknowledge their capabilities by specifying this parameter.

The only valid values are CAPABILITY_IAM, CAPABILITY_NAMED_IAM, CAPABILITY_RESOURCE_POLICY, and CAPABILITY_AUTO_EXPAND.

The following resources require you to specify CAPABILITY_IAM or CAPABILITY_NAMED_IAM: AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, and AWS::IAM::Role. If the application contains IAM resources, you can specify either CAPABILITY_IAM or CAPABILITY_NAMED_IAM. If the application contains IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

The following resources require you to specify CAPABILITY_RESOURCE_POLICY: AWS::Lambda::Permission, AWS::IAM:Policy, AWS::ApplicationAutoScaling::ScalingPolicy, AWS::S3::BucketPolicy, AWS::SQS::QueuePolicy, and AWS::SNS::TopicPolicy.

Applications that contain one or more nested applications require you to specify CAPABILITY_AUTO_EXPAND.

If your application template contains any of the above resources, we recommend that you review all permissions associated with the application before deploying. If you don't specify this parameter for an application that requires capabilities, the call will fail.

$sel:applicationId:Version', version_applicationId - The application Amazon Resource Name (ARN).

$sel:semanticVersion:Version', version_semanticVersion - The semantic version of the application:

https://semver.org/

version_sourceCodeArchiveUrl :: Lens' Version (Maybe Text)

A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

version_sourceCodeUrl :: Lens' Version (Maybe Text)

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

version_templateUrl :: Lens' Version Text

A link to the packaged AWS SAM template of your application.

version_parameterDefinitions :: Lens' Version [ParameterDefinition]

An array of parameter types supported by the application.

version_resourcesSupported :: Lens' Version Bool

Whether all of the AWS resources contained in this application are supported in the region in which it is being retrieved.

version_creationTime :: Lens' Version Text

The date and time this resource was created.

version_requiredCapabilities :: Lens' Version [Capability]

A list of values that you must specify before you can deploy certain applications. Some applications might include resources that can affect permissions in your AWS account, for example, by creating new AWS Identity and Access Management (IAM) users. For those applications, you must explicitly acknowledge their capabilities by specifying this parameter.

The only valid values are CAPABILITY_IAM, CAPABILITY_NAMED_IAM, CAPABILITY_RESOURCE_POLICY, and CAPABILITY_AUTO_EXPAND.

The following resources require you to specify CAPABILITY_IAM or CAPABILITY_NAMED_IAM: AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, and AWS::IAM::Role. If the application contains IAM resources, you can specify either CAPABILITY_IAM or CAPABILITY_NAMED_IAM. If the application contains IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

The following resources require you to specify CAPABILITY_RESOURCE_POLICY: AWS::Lambda::Permission, AWS::IAM:Policy, AWS::ApplicationAutoScaling::ScalingPolicy, AWS::S3::BucketPolicy, AWS::SQS::QueuePolicy, and AWS::SNS::TopicPolicy.

Applications that contain one or more nested applications require you to specify CAPABILITY_AUTO_EXPAND.

If your application template contains any of the above resources, we recommend that you review all permissions associated with the application before deploying. If you don't specify this parameter for an application that requires capabilities, the call will fail.

version_applicationId :: Lens' Version Text

The application Amazon Resource Name (ARN).

version_semanticVersion :: Lens' Version Text

The semantic version of the application:

https://semver.org/

VersionSummary

data VersionSummary

An application version summary.

See: newVersionSummary smart constructor.

Constructors

VersionSummary'

Fields

sourceCodeUrl :: Maybe Text

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

creationTime :: Text

The date and time this resource was created.

applicationId :: Text

The application Amazon Resource Name (ARN).

semanticVersion :: Text

The semantic version of the application:

https://semver.org/

newVersionSummary

Arguments

:: Text
$sel:creationTime:VersionSummary'

-> Text
$sel:applicationId:VersionSummary'

-> Text
$sel:semanticVersion:VersionSummary'

-> VersionSummary

Create a value of VersionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceCodeUrl:VersionSummary', versionSummary_sourceCodeUrl - A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

$sel:creationTime:VersionSummary', versionSummary_creationTime - The date and time this resource was created.

$sel:applicationId:VersionSummary', versionSummary_applicationId - The application Amazon Resource Name (ARN).

$sel:semanticVersion:VersionSummary', versionSummary_semanticVersion - The semantic version of the application:

https://semver.org/

versionSummary_sourceCodeUrl :: Lens' VersionSummary (Maybe Text)

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

versionSummary_creationTime :: Lens' VersionSummary Text

The date and time this resource was created.

versionSummary_applicationId :: Lens' VersionSummary Text

The application Amazon Resource Name (ARN).

versionSummary_semanticVersion :: Lens' VersionSummary Text

The semantic version of the application:

https://semver.org/