Network.AWS.ServerlessApplicationRepository.Types.ApplicationPolicyStatement

Description

Documentation

data ApplicationPolicyStatement

Policy statement applied to the application.

See: newApplicationPolicyStatement smart constructor.

Constructors

ApplicationPolicyStatement'

Fields

statementId :: Maybe Text

A unique ID for the statement.

principalOrgIDs :: Maybe [Text]

An array of PrinciplalOrgIDs, which corresponds to AWS IAM aws:PrincipalOrgID global condition key.

principals :: [Text]

An array of AWS account IDs, or * to make the application public.

actions :: [Text]

For the list of actions supported for this operation, see Application Permissions.

newApplicationPolicyStatement :: ApplicationPolicyStatement

Create a value of ApplicationPolicyStatement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statementId:ApplicationPolicyStatement', applicationPolicyStatement_statementId - A unique ID for the statement.

$sel:principalOrgIDs:ApplicationPolicyStatement', applicationPolicyStatement_principalOrgIDs - An array of PrinciplalOrgIDs, which corresponds to AWS IAM aws:PrincipalOrgID global condition key.

$sel:principals:ApplicationPolicyStatement', applicationPolicyStatement_principals - An array of AWS account IDs, or * to make the application public.

$sel:actions:ApplicationPolicyStatement', applicationPolicyStatement_actions - For the list of actions supported for this operation, see Application Permissions.

applicationPolicyStatement_statementId :: Lens' ApplicationPolicyStatement (Maybe Text)

A unique ID for the statement.

applicationPolicyStatement_principalOrgIDs :: Lens' ApplicationPolicyStatement (Maybe [Text])

An array of PrinciplalOrgIDs, which corresponds to AWS IAM aws:PrincipalOrgID global condition key.

applicationPolicyStatement_principals :: Lens' ApplicationPolicyStatement [Text]

An array of AWS account IDs, or * to make the application public.

applicationPolicyStatement_actions :: Lens' ApplicationPolicyStatement [Text]

For the list of actions supported for this operation, see Application Permissions.