Network.AWS.ServerlessApplicationRepository.Types.ParameterDefinition

Description

Documentation

data ParameterDefinition

Parameters supported by the application.

See: newParameterDefinition smart constructor.

Constructors

ParameterDefinition'

Fields

maxValue :: Maybe Int

A numeric value that determines the largest numeric value that you want to allow for Number types.

minLength :: Maybe Int

An integer value that determines the smallest number of characters that you want to allow for String types.

allowedValues :: Maybe [Text]

An array containing the list of values allowed for the parameter.

minValue :: Maybe Int

A numeric value that determines the smallest numeric value that you want to allow for Number types.

description :: Maybe Text

A string of up to 4,000 characters that describes the parameter.

constraintDescription :: Maybe Text

A string that explains a constraint when the constraint is violated. For example, without a constraint description, a parameter that has an allowed pattern of [A-Za-z0-9]+ displays the following error message when the user specifies an invalid value:

Malformed input-Parameter MyParameter must match pattern [A-Za-z0-9]+

By adding a constraint description, such as "must contain only uppercase and lowercase letters and numbers," you can display the following customized error message:

Malformed input-Parameter MyParameter must contain only uppercase and lowercase letters and numbers.

maxLength :: Maybe Int

An integer value that determines the largest number of characters that you want to allow for String types.

defaultValue :: Maybe Text

A value of the appropriate type for the template to use if no value is specified when a stack is created. If you define constraints for the parameter, you must specify a value that adheres to those constraints.

allowedPattern :: Maybe Text

A regular expression that represents the patterns to allow for String types.

type' :: Maybe Text

The type of the parameter.

Valid values: String | Number | List<Number> | CommaDelimitedList

String: A literal string.

For example, users can specify "MyUserName".

Number: An integer or float. AWS CloudFormation validates the parameter value as a number. However, when you use the parameter elsewhere in your template (for example, by using the Ref intrinsic function), the parameter value becomes a string.

For example, users might specify "8888".

List<Number>: An array of integers or floats that are separated by commas. AWS CloudFormation validates the parameter value as numbers. However, when you use the parameter elsewhere in your template (for example, by using the Ref intrinsic function), the parameter value becomes a list of strings.

For example, users might specify "80,20", and then Ref results in ["80","20"].

CommaDelimitedList: An array of literal strings that are separated by commas. The total number of strings should be one more than the total number of commas. Also, each member string is space-trimmed.

For example, users might specify "test,dev,prod", and then Ref results in ["test","dev","prod"].

noEcho :: Maybe Bool

Whether to mask the parameter value whenever anyone makes a call that describes the stack. If you set the value to true, the parameter value is masked with asterisks (*****).

referencedByResources :: [Text]

A list of AWS SAM resources that use this parameter.

name :: Text

The name of the parameter.

newParameterDefinition

Arguments

:: Text
$sel:name:ParameterDefinition'

-> ParameterDefinition

Create a value of ParameterDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxValue:ParameterDefinition', parameterDefinition_maxValue - A numeric value that determines the largest numeric value that you want to allow for Number types.

$sel:minLength:ParameterDefinition', parameterDefinition_minLength - An integer value that determines the smallest number of characters that you want to allow for String types.

$sel:allowedValues:ParameterDefinition', parameterDefinition_allowedValues - An array containing the list of values allowed for the parameter.

$sel:minValue:ParameterDefinition', parameterDefinition_minValue - A numeric value that determines the smallest numeric value that you want to allow for Number types.

$sel:description:ParameterDefinition', parameterDefinition_description - A string of up to 4,000 characters that describes the parameter.

$sel:constraintDescription:ParameterDefinition', parameterDefinition_constraintDescription - A string that explains a constraint when the constraint is violated. For example, without a constraint description, a parameter that has an allowed pattern of [A-Za-z0-9]+ displays the following error message when the user specifies an invalid value:

Malformed input-Parameter MyParameter must match pattern [A-Za-z0-9]+

By adding a constraint description, such as "must contain only uppercase and lowercase letters and numbers," you can display the following customized error message:

Malformed input-Parameter MyParameter must contain only uppercase and lowercase letters and numbers.

$sel:maxLength:ParameterDefinition', parameterDefinition_maxLength - An integer value that determines the largest number of characters that you want to allow for String types.

$sel:defaultValue:ParameterDefinition', parameterDefinition_defaultValue - A value of the appropriate type for the template to use if no value is specified when a stack is created. If you define constraints for the parameter, you must specify a value that adheres to those constraints.

$sel:allowedPattern:ParameterDefinition', parameterDefinition_allowedPattern - A regular expression that represents the patterns to allow for String types.

$sel:type':ParameterDefinition', parameterDefinition_type - The type of the parameter.

Valid values: String | Number | List<Number> | CommaDelimitedList

String: A literal string.

For example, users can specify "MyUserName".

Number: An integer or float. AWS CloudFormation validates the parameter value as a number. However, when you use the parameter elsewhere in your template (for example, by using the Ref intrinsic function), the parameter value becomes a string.

For example, users might specify "8888".

List<Number>: An array of integers or floats that are separated by commas. AWS CloudFormation validates the parameter value as numbers. However, when you use the parameter elsewhere in your template (for example, by using the Ref intrinsic function), the parameter value becomes a list of strings.

For example, users might specify "80,20", and then Ref results in ["80","20"].

CommaDelimitedList: An array of literal strings that are separated by commas. The total number of strings should be one more than the total number of commas. Also, each member string is space-trimmed.

For example, users might specify "test,dev,prod", and then Ref results in ["test","dev","prod"].

$sel:noEcho:ParameterDefinition', parameterDefinition_noEcho - Whether to mask the parameter value whenever anyone makes a call that describes the stack. If you set the value to true, the parameter value is masked with asterisks (*****).

$sel:referencedByResources:ParameterDefinition', parameterDefinition_referencedByResources - A list of AWS SAM resources that use this parameter.

$sel:name:ParameterDefinition', parameterDefinition_name - The name of the parameter.

parameterDefinition_maxValue :: Lens' ParameterDefinition (Maybe Int)

A numeric value that determines the largest numeric value that you want to allow for Number types.

parameterDefinition_minLength :: Lens' ParameterDefinition (Maybe Int)

An integer value that determines the smallest number of characters that you want to allow for String types.

parameterDefinition_allowedValues :: Lens' ParameterDefinition (Maybe [Text])

An array containing the list of values allowed for the parameter.

parameterDefinition_minValue :: Lens' ParameterDefinition (Maybe Int)

A numeric value that determines the smallest numeric value that you want to allow for Number types.

parameterDefinition_description :: Lens' ParameterDefinition (Maybe Text)

A string of up to 4,000 characters that describes the parameter.

parameterDefinition_constraintDescription :: Lens' ParameterDefinition (Maybe Text)

A string that explains a constraint when the constraint is violated. For example, without a constraint description, a parameter that has an allowed pattern of [A-Za-z0-9]+ displays the following error message when the user specifies an invalid value:

Malformed input-Parameter MyParameter must match pattern [A-Za-z0-9]+

By adding a constraint description, such as "must contain only uppercase and lowercase letters and numbers," you can display the following customized error message:

Malformed input-Parameter MyParameter must contain only uppercase and lowercase letters and numbers.

parameterDefinition_maxLength :: Lens' ParameterDefinition (Maybe Int)

An integer value that determines the largest number of characters that you want to allow for String types.

parameterDefinition_defaultValue :: Lens' ParameterDefinition (Maybe Text)

A value of the appropriate type for the template to use if no value is specified when a stack is created. If you define constraints for the parameter, you must specify a value that adheres to those constraints.

parameterDefinition_allowedPattern :: Lens' ParameterDefinition (Maybe Text)

A regular expression that represents the patterns to allow for String types.

parameterDefinition_type :: Lens' ParameterDefinition (Maybe Text)

The type of the parameter.

Valid values: String | Number | List<Number> | CommaDelimitedList

String: A literal string.

For example, users can specify "MyUserName".

Number: An integer or float. AWS CloudFormation validates the parameter value as a number. However, when you use the parameter elsewhere in your template (for example, by using the Ref intrinsic function), the parameter value becomes a string.

For example, users might specify "8888".

List<Number>: An array of integers or floats that are separated by commas. AWS CloudFormation validates the parameter value as numbers. However, when you use the parameter elsewhere in your template (for example, by using the Ref intrinsic function), the parameter value becomes a list of strings.

For example, users might specify "80,20", and then Ref results in ["80","20"].

CommaDelimitedList: An array of literal strings that are separated by commas. The total number of strings should be one more than the total number of commas. Also, each member string is space-trimmed.

For example, users might specify "test,dev,prod", and then Ref results in ["test","dev","prod"].

parameterDefinition_noEcho :: Lens' ParameterDefinition (Maybe Bool)

Whether to mask the parameter value whenever anyone makes a call that describes the stack. If you set the value to true, the parameter value is masked with asterisks (*****).

parameterDefinition_referencedByResources :: Lens' ParameterDefinition [Text]

A list of AWS SAM resources that use this parameter.

parameterDefinition_name :: Lens' ParameterDefinition Text

The name of the parameter.