Network.AWS.ServerlessApplicationRepository.Types.Version

Description

Documentation

data Version

Application version details.

See: newVersion smart constructor.

Constructors

Version'

Fields

sourceCodeArchiveUrl :: Maybe Text

A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

sourceCodeUrl :: Maybe Text

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

templateUrl :: Text

A link to the packaged AWS SAM template of your application.

parameterDefinitions :: [ParameterDefinition]

An array of parameter types supported by the application.

resourcesSupported :: Bool

Whether all of the AWS resources contained in this application are supported in the region in which it is being retrieved.

creationTime :: Text

The date and time this resource was created.

requiredCapabilities :: [Capability]

A list of values that you must specify before you can deploy certain applications. Some applications might include resources that can affect permissions in your AWS account, for example, by creating new AWS Identity and Access Management (IAM) users. For those applications, you must explicitly acknowledge their capabilities by specifying this parameter.

The only valid values are CAPABILITY_IAM, CAPABILITY_NAMED_IAM, CAPABILITY_RESOURCE_POLICY, and CAPABILITY_AUTO_EXPAND.

The following resources require you to specify CAPABILITY_IAM or CAPABILITY_NAMED_IAM: AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, and AWS::IAM::Role. If the application contains IAM resources, you can specify either CAPABILITY_IAM or CAPABILITY_NAMED_IAM. If the application contains IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

The following resources require you to specify CAPABILITY_RESOURCE_POLICY: AWS::Lambda::Permission, AWS::IAM:Policy, AWS::ApplicationAutoScaling::ScalingPolicy, AWS::S3::BucketPolicy, AWS::SQS::QueuePolicy, and AWS::SNS::TopicPolicy.

Applications that contain one or more nested applications require you to specify CAPABILITY_AUTO_EXPAND.

If your application template contains any of the above resources, we recommend that you review all permissions associated with the application before deploying. If you don't specify this parameter for an application that requires capabilities, the call will fail.

applicationId :: Text

The application Amazon Resource Name (ARN).

semanticVersion :: Text

The semantic version of the application:

https://semver.org/

newVersion

Arguments

:: Text
$sel:templateUrl:Version'

-> Bool
$sel:resourcesSupported:Version'

-> Text
$sel:creationTime:Version'

-> Text
$sel:applicationId:Version'

-> Text
$sel:semanticVersion:Version'

-> Version

Create a value of Version with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceCodeArchiveUrl:Version', version_sourceCodeArchiveUrl - A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

$sel:sourceCodeUrl:Version', version_sourceCodeUrl - A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

$sel:templateUrl:Version', version_templateUrl - A link to the packaged AWS SAM template of your application.

$sel:parameterDefinitions:Version', version_parameterDefinitions - An array of parameter types supported by the application.

$sel:resourcesSupported:Version', version_resourcesSupported - Whether all of the AWS resources contained in this application are supported in the region in which it is being retrieved.

$sel:creationTime:Version', version_creationTime - The date and time this resource was created.

$sel:requiredCapabilities:Version', version_requiredCapabilities - A list of values that you must specify before you can deploy certain applications. Some applications might include resources that can affect permissions in your AWS account, for example, by creating new AWS Identity and Access Management (IAM) users. For those applications, you must explicitly acknowledge their capabilities by specifying this parameter.

The only valid values are CAPABILITY_IAM, CAPABILITY_NAMED_IAM, CAPABILITY_RESOURCE_POLICY, and CAPABILITY_AUTO_EXPAND.

The following resources require you to specify CAPABILITY_IAM or CAPABILITY_NAMED_IAM: AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, and AWS::IAM::Role. If the application contains IAM resources, you can specify either CAPABILITY_IAM or CAPABILITY_NAMED_IAM. If the application contains IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

The following resources require you to specify CAPABILITY_RESOURCE_POLICY: AWS::Lambda::Permission, AWS::IAM:Policy, AWS::ApplicationAutoScaling::ScalingPolicy, AWS::S3::BucketPolicy, AWS::SQS::QueuePolicy, and AWS::SNS::TopicPolicy.

Applications that contain one or more nested applications require you to specify CAPABILITY_AUTO_EXPAND.

If your application template contains any of the above resources, we recommend that you review all permissions associated with the application before deploying. If you don't specify this parameter for an application that requires capabilities, the call will fail.

$sel:applicationId:Version', version_applicationId - The application Amazon Resource Name (ARN).

$sel:semanticVersion:Version', version_semanticVersion - The semantic version of the application:

https://semver.org/

version_sourceCodeArchiveUrl :: Lens' Version (Maybe Text)

A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

version_sourceCodeUrl :: Lens' Version (Maybe Text)

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

version_templateUrl :: Lens' Version Text

A link to the packaged AWS SAM template of your application.

version_parameterDefinitions :: Lens' Version [ParameterDefinition]

An array of parameter types supported by the application.

version_resourcesSupported :: Lens' Version Bool

Whether all of the AWS resources contained in this application are supported in the region in which it is being retrieved.

version_creationTime :: Lens' Version Text

The date and time this resource was created.

version_requiredCapabilities :: Lens' Version [Capability]

A list of values that you must specify before you can deploy certain applications. Some applications might include resources that can affect permissions in your AWS account, for example, by creating new AWS Identity and Access Management (IAM) users. For those applications, you must explicitly acknowledge their capabilities by specifying this parameter.

The only valid values are CAPABILITY_IAM, CAPABILITY_NAMED_IAM, CAPABILITY_RESOURCE_POLICY, and CAPABILITY_AUTO_EXPAND.

The following resources require you to specify CAPABILITY_IAM or CAPABILITY_NAMED_IAM: AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, and AWS::IAM::Role. If the application contains IAM resources, you can specify either CAPABILITY_IAM or CAPABILITY_NAMED_IAM. If the application contains IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

The following resources require you to specify CAPABILITY_RESOURCE_POLICY: AWS::Lambda::Permission, AWS::IAM:Policy, AWS::ApplicationAutoScaling::ScalingPolicy, AWS::S3::BucketPolicy, AWS::SQS::QueuePolicy, and AWS::SNS::TopicPolicy.

Applications that contain one or more nested applications require you to specify CAPABILITY_AUTO_EXPAND.

If your application template contains any of the above resources, we recommend that you review all permissions associated with the application before deploying. If you don't specify this parameter for an application that requires capabilities, the call will fail.

version_applicationId :: Lens' Version Text

The application Amazon Resource Name (ARN).

version_semanticVersion :: Lens' Version Text

The semantic version of the application:

https://semver.org/