Network.AWS.ServerlessApplicationRepository.UpdateApplication

Description

Updates the specified application.

Creating a Request

data UpdateApplication

See: newUpdateApplication smart constructor.

Constructors

UpdateApplication'

Fields

homePageUrl :: Maybe Text

A URL with more information about the application, for example the location of your GitHub repository for the application.

readmeBody :: Maybe Text

A text readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

readmeUrl :: Maybe Text

A link to the readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

author :: Maybe Text

The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

labels :: Maybe [Text]

Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

description :: Maybe Text

The description of the application.

Minimum length=1. Maximum length=256

applicationId :: Text

The Amazon Resource Name (ARN) of the application.

newUpdateApplication

Arguments

:: Text
$sel:applicationId:UpdateApplication'

-> UpdateApplication

Create a value of UpdateApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:homePageUrl:UpdateApplication', updateApplication_homePageUrl - A URL with more information about the application, for example the location of your GitHub repository for the application.

$sel:readmeBody:UpdateApplication', updateApplication_readmeBody - A text readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

$sel:readmeUrl:UpdateApplication', updateApplication_readmeUrl - A link to the readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

$sel:author:UpdateApplication', updateApplication_author - The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

$sel:labels:UpdateApplication', updateApplication_labels - Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

$sel:description:UpdateApplication', updateApplication_description - The description of the application.

Minimum length=1. Maximum length=256

$sel:applicationId:UpdateApplication', updateApplication_applicationId - The Amazon Resource Name (ARN) of the application.

Request Lenses

updateApplication_homePageUrl :: Lens' UpdateApplication (Maybe Text)

A URL with more information about the application, for example the location of your GitHub repository for the application.

updateApplication_readmeBody :: Lens' UpdateApplication (Maybe Text)

A text readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

updateApplication_readmeUrl :: Lens' UpdateApplication (Maybe Text)

A link to the readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

updateApplication_author :: Lens' UpdateApplication (Maybe Text)

The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

updateApplication_labels :: Lens' UpdateApplication (Maybe [Text])

Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

updateApplication_description :: Lens' UpdateApplication (Maybe Text)

The description of the application.

Minimum length=1. Maximum length=256

updateApplication_applicationId :: Lens' UpdateApplication Text

The Amazon Resource Name (ARN) of the application.

Destructuring the Response

data UpdateApplicationResponse

See: newUpdateApplicationResponse smart constructor.

Constructors

UpdateApplicationResponse'

Fields

creationTime :: Maybe Text

The date and time this resource was created.

homePageUrl :: Maybe Text

A URL with more information about the application, for example the location of your GitHub repository for the application.

licenseUrl :: Maybe Text

A link to a license file of the app that matches the spdxLicenseID value of your application.

Maximum size 5 MB

readmeUrl :: Maybe Text

A link to the readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

applicationId :: Maybe Text

The application Amazon Resource Name (ARN).

name :: Maybe Text

The name of the application.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

version :: Maybe Version

Version information about the application.

author :: Maybe Text

The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

labels :: Maybe [Text]

Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

verifiedAuthorUrl :: Maybe Text

The URL to the public profile of a verified author. This URL is submitted by the author.

description :: Maybe Text

The description of the application.

Minimum length=1. Maximum length=256

spdxLicenseId :: Maybe Text

A valid identifier from https://spdx.org/licenses/.

isVerifiedAuthor :: Maybe Bool

Whether the author of this application has been verified. This means means that AWS has made a good faith review, as a reasonable and prudent service provider, of the information provided by the requester and has confirmed that the requester's identity is as claimed.

httpStatus :: Int

The response's http status code.

newUpdateApplicationResponse

Arguments

:: Int
$sel:httpStatus:UpdateApplicationResponse'

-> UpdateApplicationResponse

Create a value of UpdateApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:UpdateApplicationResponse', updateApplicationResponse_creationTime - The date and time this resource was created.

$sel:homePageUrl:UpdateApplication', updateApplicationResponse_homePageUrl - A URL with more information about the application, for example the location of your GitHub repository for the application.

$sel:licenseUrl:UpdateApplicationResponse', updateApplicationResponse_licenseUrl - A link to a license file of the app that matches the spdxLicenseID value of your application.

Maximum size 5 MB

$sel:readmeUrl:UpdateApplication', updateApplicationResponse_readmeUrl - A link to the readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

$sel:applicationId:UpdateApplication', updateApplicationResponse_applicationId - The application Amazon Resource Name (ARN).

$sel:name:UpdateApplicationResponse', updateApplicationResponse_name - The name of the application.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

$sel:version:UpdateApplicationResponse', updateApplicationResponse_version - Version information about the application.

$sel:author:UpdateApplication', updateApplicationResponse_author - The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

$sel:labels:UpdateApplication', updateApplicationResponse_labels - Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

$sel:verifiedAuthorUrl:UpdateApplicationResponse', updateApplicationResponse_verifiedAuthorUrl - The URL to the public profile of a verified author. This URL is submitted by the author.

$sel:description:UpdateApplication', updateApplicationResponse_description - The description of the application.

Minimum length=1. Maximum length=256

$sel:spdxLicenseId:UpdateApplicationResponse', updateApplicationResponse_spdxLicenseId - A valid identifier from https://spdx.org/licenses/.

$sel:isVerifiedAuthor:UpdateApplicationResponse', updateApplicationResponse_isVerifiedAuthor - Whether the author of this application has been verified. This means means that AWS has made a good faith review, as a reasonable and prudent service provider, of the information provided by the requester and has confirmed that the requester's identity is as claimed.

$sel:httpStatus:UpdateApplicationResponse', updateApplicationResponse_httpStatus - The response's http status code.

Response Lenses

updateApplicationResponse_creationTime :: Lens' UpdateApplicationResponse (Maybe Text)

The date and time this resource was created.

updateApplicationResponse_homePageUrl :: Lens' UpdateApplicationResponse (Maybe Text)

A URL with more information about the application, for example the location of your GitHub repository for the application.

updateApplicationResponse_licenseUrl :: Lens' UpdateApplicationResponse (Maybe Text)

A link to a license file of the app that matches the spdxLicenseID value of your application.

Maximum size 5 MB

updateApplicationResponse_readmeUrl :: Lens' UpdateApplicationResponse (Maybe Text)

A link to the readme file in Markdown language that contains a more detailed description of the application and how it works.

Maximum size 5 MB

updateApplicationResponse_applicationId :: Lens' UpdateApplicationResponse (Maybe Text)

The application Amazon Resource Name (ARN).

updateApplicationResponse_name :: Lens' UpdateApplicationResponse (Maybe Text)

The name of the application.

Minimum length=1. Maximum length=140

Pattern: "[a-zA-Z0-9\\-]+";

updateApplicationResponse_version :: Lens' UpdateApplicationResponse (Maybe Version)

Version information about the application.

updateApplicationResponse_author :: Lens' UpdateApplicationResponse (Maybe Text)

The name of the author publishing the app.

Minimum length=1. Maximum length=127.

Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";

updateApplicationResponse_labels :: Lens' UpdateApplicationResponse (Maybe [Text])

Labels to improve discovery of apps in search results.

Minimum length=1. Maximum length=127. Maximum number of labels: 10

Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";

updateApplicationResponse_verifiedAuthorUrl :: Lens' UpdateApplicationResponse (Maybe Text)

The URL to the public profile of a verified author. This URL is submitted by the author.

updateApplicationResponse_description :: Lens' UpdateApplicationResponse (Maybe Text)

The description of the application.

Minimum length=1. Maximum length=256

updateApplicationResponse_spdxLicenseId :: Lens' UpdateApplicationResponse (Maybe Text)

A valid identifier from https://spdx.org/licenses/.

updateApplicationResponse_isVerifiedAuthor :: Lens' UpdateApplicationResponse (Maybe Bool)

Whether the author of this application has been verified. This means means that AWS has made a good faith review, as a reasonable and prudent service provider, of the information provided by the requester and has confirmed that the requester's identity is as claimed.

updateApplicationResponse_httpStatus :: Lens' UpdateApplicationResponse Int

The response's http status code.