Network.AWS.ServiceCatalog.AssociateServiceActionWithProvisioningArtifact

Description

Associates a self-service action with a provisioning artifact.

Creating a Request

data AssociateServiceActionWithProvisioningArtifact

See: newAssociateServiceActionWithProvisioningArtifact smart constructor.

Constructors

AssociateServiceActionWithProvisioningArtifact'

Fields

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

productId :: Text

The product identifier. For example, prod-abcdzk7xy33qa.

provisioningArtifactId :: Text

The identifier of the provisioning artifact. For example, pa-4abcdjnxjj6ne.

serviceActionId :: Text

The self-service action identifier. For example, act-fs7abcd89wxyz.

newAssociateServiceActionWithProvisioningArtifact

Arguments

:: Text
$sel:productId:AssociateServiceActionWithProvisioningArtifact'

-> Text
$sel:provisioningArtifactId:AssociateServiceActionWithProvisioningArtifact'

-> Text
$sel:serviceActionId:AssociateServiceActionWithProvisioningArtifact'

-> AssociateServiceActionWithProvisioningArtifact

Create a value of AssociateServiceActionWithProvisioningArtifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceptLanguage:AssociateServiceActionWithProvisioningArtifact', associateServiceActionWithProvisioningArtifact_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:productId:AssociateServiceActionWithProvisioningArtifact', associateServiceActionWithProvisioningArtifact_productId - The product identifier. For example, prod-abcdzk7xy33qa.

$sel:provisioningArtifactId:AssociateServiceActionWithProvisioningArtifact', associateServiceActionWithProvisioningArtifact_provisioningArtifactId - The identifier of the provisioning artifact. For example, pa-4abcdjnxjj6ne.

$sel:serviceActionId:AssociateServiceActionWithProvisioningArtifact', associateServiceActionWithProvisioningArtifact_serviceActionId - The self-service action identifier. For example, act-fs7abcd89wxyz.

Request Lenses

associateServiceActionWithProvisioningArtifact_acceptLanguage :: Lens' AssociateServiceActionWithProvisioningArtifact (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

associateServiceActionWithProvisioningArtifact_productId :: Lens' AssociateServiceActionWithProvisioningArtifact Text

The product identifier. For example, prod-abcdzk7xy33qa.

associateServiceActionWithProvisioningArtifact_provisioningArtifactId :: Lens' AssociateServiceActionWithProvisioningArtifact Text

The identifier of the provisioning artifact. For example, pa-4abcdjnxjj6ne.

associateServiceActionWithProvisioningArtifact_serviceActionId :: Lens' AssociateServiceActionWithProvisioningArtifact Text

The self-service action identifier. For example, act-fs7abcd89wxyz.

Destructuring the Response

data AssociateServiceActionWithProvisioningArtifactResponse

See: newAssociateServiceActionWithProvisioningArtifactResponse smart constructor.

Constructors

AssociateServiceActionWithProvisioningArtifactResponse'

Fields

httpStatus :: Int

The response's http status code.

newAssociateServiceActionWithProvisioningArtifactResponse

Arguments

:: Int
$sel:httpStatus:AssociateServiceActionWithProvisioningArtifactResponse'

-> AssociateServiceActionWithProvisioningArtifactResponse

Create a value of AssociateServiceActionWithProvisioningArtifactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AssociateServiceActionWithProvisioningArtifactResponse', associateServiceActionWithProvisioningArtifactResponse_httpStatus - The response's http status code.

Response Lenses

associateServiceActionWithProvisioningArtifactResponse_httpStatus :: Lens' AssociateServiceActionWithProvisioningArtifactResponse Int

The response's http status code.