Network.AWS.ServiceCatalog.CopyProduct

Description

Copies the specified source product to the specified target product or a new product.

You can copy a product to the same account or another account. You can copy a product to the same region or another region.

This operation is performed asynchronously. To track the progress of the operation, use DescribeCopyProductStatus.

Creating a Request

data CopyProduct

See: newCopyProduct smart constructor.

Constructors

CopyProduct'

Fields

copyOptions :: Maybe [CopyOption]

The copy options. If the value is CopyTags, the tags from the source product are copied to the target product.

targetProductName :: Maybe Text

A name for the target product. The default is the name of the source product.

targetProductId :: Maybe Text

The identifier of the target product. By default, a new product is created.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

sourceProvisioningArtifactIdentifiers :: Maybe [HashMap ProvisioningArtifactPropertyName Text]

The identifiers of the provisioning artifacts (also known as versions) of the product to copy. By default, all provisioning artifacts are copied.

sourceProductArn :: Text

The Amazon Resource Name (ARN) of the source product.

idempotencyToken :: Text

A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

newCopyProduct

Arguments

:: Text
$sel:sourceProductArn:CopyProduct'

-> Text
$sel:idempotencyToken:CopyProduct'

-> CopyProduct

Create a value of CopyProduct with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:copyOptions:CopyProduct', copyProduct_copyOptions - The copy options. If the value is CopyTags, the tags from the source product are copied to the target product.

$sel:targetProductName:CopyProduct', copyProduct_targetProductName - A name for the target product. The default is the name of the source product.

$sel:targetProductId:CopyProduct', copyProduct_targetProductId - The identifier of the target product. By default, a new product is created.

$sel:acceptLanguage:CopyProduct', copyProduct_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:sourceProvisioningArtifactIdentifiers:CopyProduct', copyProduct_sourceProvisioningArtifactIdentifiers - The identifiers of the provisioning artifacts (also known as versions) of the product to copy. By default, all provisioning artifacts are copied.

$sel:sourceProductArn:CopyProduct', copyProduct_sourceProductArn - The Amazon Resource Name (ARN) of the source product.

$sel:idempotencyToken:CopyProduct', copyProduct_idempotencyToken - A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

Request Lenses

copyProduct_copyOptions :: Lens' CopyProduct (Maybe [CopyOption])

The copy options. If the value is CopyTags, the tags from the source product are copied to the target product.

copyProduct_targetProductName :: Lens' CopyProduct (Maybe Text)

A name for the target product. The default is the name of the source product.

copyProduct_targetProductId :: Lens' CopyProduct (Maybe Text)

The identifier of the target product. By default, a new product is created.

copyProduct_acceptLanguage :: Lens' CopyProduct (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

copyProduct_sourceProvisioningArtifactIdentifiers :: Lens' CopyProduct (Maybe [HashMap ProvisioningArtifactPropertyName Text])

The identifiers of the provisioning artifacts (also known as versions) of the product to copy. By default, all provisioning artifacts are copied.

copyProduct_sourceProductArn :: Lens' CopyProduct Text

The Amazon Resource Name (ARN) of the source product.

copyProduct_idempotencyToken :: Lens' CopyProduct Text

A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

Destructuring the Response

data CopyProductResponse

See: newCopyProductResponse smart constructor.

Constructors

CopyProductResponse'

Fields

copyProductToken :: Maybe Text

The token to use to track the progress of the operation.

httpStatus :: Int

The response's http status code.

newCopyProductResponse

Arguments

:: Int
$sel:httpStatus:CopyProductResponse'

-> CopyProductResponse

Create a value of CopyProductResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:copyProductToken:CopyProductResponse', copyProductResponse_copyProductToken - The token to use to track the progress of the operation.

$sel:httpStatus:CopyProductResponse', copyProductResponse_httpStatus - The response's http status code.

Response Lenses

copyProductResponse_copyProductToken :: Lens' CopyProductResponse (Maybe Text)

The token to use to track the progress of the operation.

copyProductResponse_httpStatus :: Lens' CopyProductResponse Int

The response's http status code.