Network.AWS.ServiceCatalog.CreateProduct

Description

Creates a product.

A delegated admin is authorized to invoke this command.

The user or role that performs this operation must have the cloudformation:GetTemplate IAM policy permission. This policy permission is required when using the ImportFromPhysicalId template source in the information data section.

Creating a Request

data CreateProduct

See: newCreateProduct smart constructor.

Constructors

CreateProduct'

Fields

distributor :: Maybe Text

The distributor of the product.

supportUrl :: Maybe Text

The contact URL for product support.

^https?:\/\// / is the pattern used to validate SupportUrl.

tags :: Maybe [Tag]

One or more tags.

supportDescription :: Maybe Text

The support information about the product.

description :: Maybe Text

The description of the product.

supportEmail :: Maybe Text

The contact email for product support.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

name :: Text

The name of the product.

owner :: Text

The owner of the product.

productType :: ProductType

The type of product.

provisioningArtifactParameters :: ProvisioningArtifactProperties

The configuration of the provisioning artifact.

idempotencyToken :: Text

A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

newCreateProduct

Arguments

:: Text
$sel:name:CreateProduct'

-> Text
$sel:owner:CreateProduct'

-> ProductType
$sel:productType:CreateProduct'

-> ProvisioningArtifactProperties
$sel:provisioningArtifactParameters:CreateProduct'

-> Text
$sel:idempotencyToken:CreateProduct'

-> CreateProduct

Create a value of CreateProduct with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:distributor:CreateProduct', createProduct_distributor - The distributor of the product.

$sel:supportUrl:CreateProduct', createProduct_supportUrl - The contact URL for product support.

^https?:\/\// / is the pattern used to validate SupportUrl.

$sel:tags:CreateProduct', createProduct_tags - One or more tags.

$sel:supportDescription:CreateProduct', createProduct_supportDescription - The support information about the product.

$sel:description:CreateProduct', createProduct_description - The description of the product.

$sel:supportEmail:CreateProduct', createProduct_supportEmail - The contact email for product support.

$sel:acceptLanguage:CreateProduct', createProduct_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:name:CreateProduct', createProduct_name - The name of the product.

$sel:owner:CreateProduct', createProduct_owner - The owner of the product.

$sel:productType:CreateProduct', createProduct_productType - The type of product.

$sel:provisioningArtifactParameters:CreateProduct', createProduct_provisioningArtifactParameters - The configuration of the provisioning artifact.

$sel:idempotencyToken:CreateProduct', createProduct_idempotencyToken - A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

Request Lenses

createProduct_distributor :: Lens' CreateProduct (Maybe Text)

The distributor of the product.

createProduct_supportUrl :: Lens' CreateProduct (Maybe Text)

The contact URL for product support.

^https?:\/\// / is the pattern used to validate SupportUrl.

createProduct_tags :: Lens' CreateProduct (Maybe [Tag])

One or more tags.

createProduct_supportDescription :: Lens' CreateProduct (Maybe Text)

The support information about the product.

createProduct_description :: Lens' CreateProduct (Maybe Text)

The description of the product.

createProduct_supportEmail :: Lens' CreateProduct (Maybe Text)

The contact email for product support.

createProduct_acceptLanguage :: Lens' CreateProduct (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

createProduct_name :: Lens' CreateProduct Text

The name of the product.

createProduct_owner :: Lens' CreateProduct Text

The owner of the product.

createProduct_productType :: Lens' CreateProduct ProductType

The type of product.

createProduct_provisioningArtifactParameters :: Lens' CreateProduct ProvisioningArtifactProperties

The configuration of the provisioning artifact.

createProduct_idempotencyToken :: Lens' CreateProduct Text

A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

Destructuring the Response

data CreateProductResponse

See: newCreateProductResponse smart constructor.

Constructors

CreateProductResponse'

Fields

productViewDetail :: Maybe ProductViewDetail

Information about the product view.

provisioningArtifactDetail :: Maybe ProvisioningArtifactDetail

Information about the provisioning artifact.

tags :: Maybe [Tag]

Information about the tags associated with the product.

httpStatus :: Int

The response's http status code.

newCreateProductResponse

Arguments

:: Int
$sel:httpStatus:CreateProductResponse'

-> CreateProductResponse

Create a value of CreateProductResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:productViewDetail:CreateProductResponse', createProductResponse_productViewDetail - Information about the product view.

$sel:provisioningArtifactDetail:CreateProductResponse', createProductResponse_provisioningArtifactDetail - Information about the provisioning artifact.

$sel:tags:CreateProduct', createProductResponse_tags - Information about the tags associated with the product.

$sel:httpStatus:CreateProductResponse', createProductResponse_httpStatus - The response's http status code.

Response Lenses

createProductResponse_productViewDetail :: Lens' CreateProductResponse (Maybe ProductViewDetail)

Information about the product view.

createProductResponse_provisioningArtifactDetail :: Lens' CreateProductResponse (Maybe ProvisioningArtifactDetail)

Information about the provisioning artifact.

createProductResponse_tags :: Lens' CreateProductResponse (Maybe [Tag])

Information about the tags associated with the product.

createProductResponse_httpStatus :: Lens' CreateProductResponse Int

The response's http status code.