Network.AWS.ServiceCatalog.CreateProvisionedProductPlan

Description

Creates a plan. A plan includes the list of resources to be created (when provisioning a new product) or modified (when updating a provisioned product) when the plan is executed.

You can create one plan per provisioned product. To create a plan for an existing provisioned product, the product status must be AVAILBLE or TAINTED.

To view the resource changes in the change set, use DescribeProvisionedProductPlan. To create or modify the provisioned product, use ExecuteProvisionedProductPlan.

Creating a Request

data CreateProvisionedProductPlan

See: newCreateProvisionedProductPlan smart constructor.

Constructors

CreateProvisionedProductPlan'

Fields

notificationArns :: Maybe [Text]

Passed to CloudFormation. The SNS topic ARNs to which to publish stack-related events.

tags :: Maybe [Tag]

One or more tags.

If the plan is for an existing provisioned product, the product must have a RESOURCE_UPDATE constraint with TagUpdatesOnProvisionedProduct set to ALLOWED to allow tag updates.

provisioningParameters :: Maybe [UpdateProvisioningParameter]

Parameters specified by the administrator that are required for provisioning the product.

pathId :: Maybe Text

The path identifier of the product. This value is optional if the product has a default path, and required if the product has more than one path. To list the paths for a product, use ListLaunchPaths.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

planName :: Text

The name of the plan.

planType :: ProvisionedProductPlanType

The plan type.

productId :: Text

The product identifier.

provisionedProductName :: Text

A user-friendly name for the provisioned product. This value must be unique for the AWS account and cannot be updated after the product is provisioned.

provisioningArtifactId :: Text

The identifier of the provisioning artifact.

idempotencyToken :: Text

A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

newCreateProvisionedProductPlan

Arguments

:: Text
$sel:planName:CreateProvisionedProductPlan'

-> ProvisionedProductPlanType
$sel:planType:CreateProvisionedProductPlan'

-> Text
$sel:productId:CreateProvisionedProductPlan'

-> Text
$sel:provisionedProductName:CreateProvisionedProductPlan'

-> Text
$sel:provisioningArtifactId:CreateProvisionedProductPlan'

-> Text
$sel:idempotencyToken:CreateProvisionedProductPlan'

-> CreateProvisionedProductPlan

Create a value of CreateProvisionedProductPlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notificationArns:CreateProvisionedProductPlan', createProvisionedProductPlan_notificationArns - Passed to CloudFormation. The SNS topic ARNs to which to publish stack-related events.

$sel:tags:CreateProvisionedProductPlan', createProvisionedProductPlan_tags - One or more tags.

If the plan is for an existing provisioned product, the product must have a RESOURCE_UPDATE constraint with TagUpdatesOnProvisionedProduct set to ALLOWED to allow tag updates.

$sel:provisioningParameters:CreateProvisionedProductPlan', createProvisionedProductPlan_provisioningParameters - Parameters specified by the administrator that are required for provisioning the product.

$sel:pathId:CreateProvisionedProductPlan', createProvisionedProductPlan_pathId - The path identifier of the product. This value is optional if the product has a default path, and required if the product has more than one path. To list the paths for a product, use ListLaunchPaths.

$sel:acceptLanguage:CreateProvisionedProductPlan', createProvisionedProductPlan_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:planName:CreateProvisionedProductPlan', createProvisionedProductPlan_planName - The name of the plan.

$sel:planType:CreateProvisionedProductPlan', createProvisionedProductPlan_planType - The plan type.

$sel:productId:CreateProvisionedProductPlan', createProvisionedProductPlan_productId - The product identifier.

$sel:provisionedProductName:CreateProvisionedProductPlan', createProvisionedProductPlan_provisionedProductName - A user-friendly name for the provisioned product. This value must be unique for the AWS account and cannot be updated after the product is provisioned.

$sel:provisioningArtifactId:CreateProvisionedProductPlan', createProvisionedProductPlan_provisioningArtifactId - The identifier of the provisioning artifact.

$sel:idempotencyToken:CreateProvisionedProductPlan', createProvisionedProductPlan_idempotencyToken - A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

Request Lenses

createProvisionedProductPlan_notificationArns :: Lens' CreateProvisionedProductPlan (Maybe [Text])

Passed to CloudFormation. The SNS topic ARNs to which to publish stack-related events.

createProvisionedProductPlan_tags :: Lens' CreateProvisionedProductPlan (Maybe [Tag])

One or more tags.

If the plan is for an existing provisioned product, the product must have a RESOURCE_UPDATE constraint with TagUpdatesOnProvisionedProduct set to ALLOWED to allow tag updates.

createProvisionedProductPlan_provisioningParameters :: Lens' CreateProvisionedProductPlan (Maybe [UpdateProvisioningParameter])

Parameters specified by the administrator that are required for provisioning the product.

createProvisionedProductPlan_pathId :: Lens' CreateProvisionedProductPlan (Maybe Text)

The path identifier of the product. This value is optional if the product has a default path, and required if the product has more than one path. To list the paths for a product, use ListLaunchPaths.

createProvisionedProductPlan_acceptLanguage :: Lens' CreateProvisionedProductPlan (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

createProvisionedProductPlan_planName :: Lens' CreateProvisionedProductPlan Text

The name of the plan.

createProvisionedProductPlan_planType :: Lens' CreateProvisionedProductPlan ProvisionedProductPlanType

The plan type.

createProvisionedProductPlan_productId :: Lens' CreateProvisionedProductPlan Text

The product identifier.

createProvisionedProductPlan_provisionedProductName :: Lens' CreateProvisionedProductPlan Text

A user-friendly name for the provisioned product. This value must be unique for the AWS account and cannot be updated after the product is provisioned.

createProvisionedProductPlan_provisioningArtifactId :: Lens' CreateProvisionedProductPlan Text

The identifier of the provisioning artifact.

createProvisionedProductPlan_idempotencyToken :: Lens' CreateProvisionedProductPlan Text

A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

Destructuring the Response

data CreateProvisionedProductPlanResponse

See: newCreateProvisionedProductPlanResponse smart constructor.

Constructors

CreateProvisionedProductPlanResponse'

Fields

provisionProductId :: Maybe Text

The product identifier.

provisionedProductName :: Maybe Text

The user-friendly name of the provisioned product.

provisioningArtifactId :: Maybe Text

The identifier of the provisioning artifact.

planName :: Maybe Text

The name of the plan.

planId :: Maybe Text

The plan identifier.

httpStatus :: Int

The response's http status code.

newCreateProvisionedProductPlanResponse

Arguments

:: Int
$sel:httpStatus:CreateProvisionedProductPlanResponse'

-> CreateProvisionedProductPlanResponse

Create a value of CreateProvisionedProductPlanResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:provisionProductId:CreateProvisionedProductPlanResponse', createProvisionedProductPlanResponse_provisionProductId - The product identifier.

$sel:provisionedProductName:CreateProvisionedProductPlan', createProvisionedProductPlanResponse_provisionedProductName - The user-friendly name of the provisioned product.

$sel:provisioningArtifactId:CreateProvisionedProductPlan', createProvisionedProductPlanResponse_provisioningArtifactId - The identifier of the provisioning artifact.

$sel:planName:CreateProvisionedProductPlan', createProvisionedProductPlanResponse_planName - The name of the plan.

$sel:planId:CreateProvisionedProductPlanResponse', createProvisionedProductPlanResponse_planId - The plan identifier.

$sel:httpStatus:CreateProvisionedProductPlanResponse', createProvisionedProductPlanResponse_httpStatus - The response's http status code.

Response Lenses

createProvisionedProductPlanResponse_provisionProductId :: Lens' CreateProvisionedProductPlanResponse (Maybe Text)

The product identifier.

createProvisionedProductPlanResponse_provisionedProductName :: Lens' CreateProvisionedProductPlanResponse (Maybe Text)

The user-friendly name of the provisioned product.

createProvisionedProductPlanResponse_provisioningArtifactId :: Lens' CreateProvisionedProductPlanResponse (Maybe Text)

The identifier of the provisioning artifact.

createProvisionedProductPlanResponse_planName :: Lens' CreateProvisionedProductPlanResponse (Maybe Text)

The name of the plan.

createProvisionedProductPlanResponse_planId :: Lens' CreateProvisionedProductPlanResponse (Maybe Text)

The plan identifier.

createProvisionedProductPlanResponse_httpStatus :: Lens' CreateProvisionedProductPlanResponse Int

The response's http status code.