Network.AWS.ServiceCatalog.CreateServiceAction

Description

Creates a self-service action.

Creating a Request

data CreateServiceAction

See: newCreateServiceAction smart constructor.

Constructors

CreateServiceAction'

Fields

description :: Maybe Text

The self-service action description.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

name :: Text

The self-service action name.

definitionType :: ServiceActionDefinitionType

The service action definition type. For example, SSM_AUTOMATION.

definition :: HashMap ServiceActionDefinitionKey Text

The self-service action definition. Can be one of the following:

Name

The name of the AWS Systems Manager document (SSM document). For example, AWS-RestartEC2Instance.
If you are using a shared SSM document, you must provide the ARN instead of the name.

Version

The AWS Systems Manager automation document version. For example, "Version": "1"

AssumeRole

The Amazon Resource Name (ARN) of the role that performs the self-service actions on your behalf. For example, "AssumeRole": "arn:aws:iam::12345678910:role/ActionRole".
To reuse the provisioned product launch role, set to "AssumeRole": "LAUNCH_ROLE".

Parameters

The list of parameters in JSON format.
For example: [{\"Name\":\"InstanceId\",\"Type\":\"TARGET\"}] or [{\"Name\":\"InstanceId\",\"Type\":\"TEXT_VALUE\"}].

idempotencyToken :: Text

A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

newCreateServiceAction

Arguments

:: Text
$sel:name:CreateServiceAction'

-> ServiceActionDefinitionType
$sel:definitionType:CreateServiceAction'

-> Text
$sel:idempotencyToken:CreateServiceAction'

-> CreateServiceAction

Create a value of CreateServiceAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateServiceAction', createServiceAction_description - The self-service action description.

$sel:acceptLanguage:CreateServiceAction', createServiceAction_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:name:CreateServiceAction', createServiceAction_name - The self-service action name.

$sel:definitionType:CreateServiceAction', createServiceAction_definitionType - The service action definition type. For example, SSM_AUTOMATION.

$sel:definition:CreateServiceAction', createServiceAction_definition - The self-service action definition. Can be one of the following:

Name

The name of the AWS Systems Manager document (SSM document). For example, AWS-RestartEC2Instance.
If you are using a shared SSM document, you must provide the ARN instead of the name.

Version

The AWS Systems Manager automation document version. For example, "Version": "1"

AssumeRole

The Amazon Resource Name (ARN) of the role that performs the self-service actions on your behalf. For example, "AssumeRole": "arn:aws:iam::12345678910:role/ActionRole".
To reuse the provisioned product launch role, set to "AssumeRole": "LAUNCH_ROLE".

Parameters

The list of parameters in JSON format.
For example: [{\"Name\":\"InstanceId\",\"Type\":\"TARGET\"}] or [{\"Name\":\"InstanceId\",\"Type\":\"TEXT_VALUE\"}].

$sel:idempotencyToken:CreateServiceAction', createServiceAction_idempotencyToken - A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

Request Lenses

createServiceAction_description :: Lens' CreateServiceAction (Maybe Text)

The self-service action description.

createServiceAction_acceptLanguage :: Lens' CreateServiceAction (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

createServiceAction_name :: Lens' CreateServiceAction Text

The self-service action name.

createServiceAction_definitionType :: Lens' CreateServiceAction ServiceActionDefinitionType

The service action definition type. For example, SSM_AUTOMATION.

createServiceAction_definition :: Lens' CreateServiceAction (HashMap ServiceActionDefinitionKey Text)

The self-service action definition. Can be one of the following:

Name

The name of the AWS Systems Manager document (SSM document). For example, AWS-RestartEC2Instance.
If you are using a shared SSM document, you must provide the ARN instead of the name.

Version

The AWS Systems Manager automation document version. For example, "Version": "1"

AssumeRole

The Amazon Resource Name (ARN) of the role that performs the self-service actions on your behalf. For example, "AssumeRole": "arn:aws:iam::12345678910:role/ActionRole".
To reuse the provisioned product launch role, set to "AssumeRole": "LAUNCH_ROLE".

Parameters

The list of parameters in JSON format.
For example: [{\"Name\":\"InstanceId\",\"Type\":\"TARGET\"}] or [{\"Name\":\"InstanceId\",\"Type\":\"TEXT_VALUE\"}].

createServiceAction_idempotencyToken :: Lens' CreateServiceAction Text

A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

Destructuring the Response

data CreateServiceActionResponse

See: newCreateServiceActionResponse smart constructor.

Constructors

CreateServiceActionResponse'

Fields

serviceActionDetail :: Maybe ServiceActionDetail

An object containing information about the self-service action.

httpStatus :: Int

The response's http status code.

newCreateServiceActionResponse

Arguments

:: Int
$sel:httpStatus:CreateServiceActionResponse'

-> CreateServiceActionResponse

Create a value of CreateServiceActionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceActionDetail:CreateServiceActionResponse', createServiceActionResponse_serviceActionDetail - An object containing information about the self-service action.

$sel:httpStatus:CreateServiceActionResponse', createServiceActionResponse_httpStatus - The response's http status code.

Response Lenses

createServiceActionResponse_serviceActionDetail :: Lens' CreateServiceActionResponse (Maybe ServiceActionDetail)

An object containing information about the self-service action.

createServiceActionResponse_httpStatus :: Lens' CreateServiceActionResponse Int

The response's http status code.