Network.AWS.ServiceCatalog.DeleteServiceAction

Description

Deletes a self-service action.

Creating a Request

data DeleteServiceAction

See: newDeleteServiceAction smart constructor.

Constructors

DeleteServiceAction'

Fields

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

id :: Text

The self-service action identifier. For example, act-fs7abcd89wxyz.

newDeleteServiceAction

Arguments

:: Text
$sel:id:DeleteServiceAction'

-> DeleteServiceAction

Create a value of DeleteServiceAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceptLanguage:DeleteServiceAction', deleteServiceAction_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:id:DeleteServiceAction', deleteServiceAction_id - The self-service action identifier. For example, act-fs7abcd89wxyz.

Request Lenses

deleteServiceAction_acceptLanguage :: Lens' DeleteServiceAction (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

deleteServiceAction_id :: Lens' DeleteServiceAction Text

The self-service action identifier. For example, act-fs7abcd89wxyz.

Destructuring the Response

data DeleteServiceActionResponse

See: newDeleteServiceActionResponse smart constructor.

Constructors

DeleteServiceActionResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteServiceActionResponse

Arguments

:: Int
$sel:httpStatus:DeleteServiceActionResponse'

-> DeleteServiceActionResponse

Create a value of DeleteServiceActionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteServiceActionResponse', deleteServiceActionResponse_httpStatus - The response's http status code.

Response Lenses

deleteServiceActionResponse_httpStatus :: Lens' DeleteServiceActionResponse Int

The response's http status code.