Network.AWS.ServiceCatalog.DescribeCopyProductStatus

Description

Gets the status of the specified copy product operation.

Creating a Request

data DescribeCopyProductStatus

See: newDescribeCopyProductStatus smart constructor.

Constructors

DescribeCopyProductStatus'

Fields

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

copyProductToken :: Text

The token for the copy product operation. This token is returned by CopyProduct.

newDescribeCopyProductStatus

Arguments

:: Text
$sel:copyProductToken:DescribeCopyProductStatus'

-> DescribeCopyProductStatus

Create a value of DescribeCopyProductStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceptLanguage:DescribeCopyProductStatus', describeCopyProductStatus_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:copyProductToken:DescribeCopyProductStatus', describeCopyProductStatus_copyProductToken - The token for the copy product operation. This token is returned by CopyProduct.

Request Lenses

describeCopyProductStatus_acceptLanguage :: Lens' DescribeCopyProductStatus (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

describeCopyProductStatus_copyProductToken :: Lens' DescribeCopyProductStatus Text

The token for the copy product operation. This token is returned by CopyProduct.

Destructuring the Response

data DescribeCopyProductStatusResponse

See: newDescribeCopyProductStatusResponse smart constructor.

Constructors

DescribeCopyProductStatusResponse'

Fields

statusDetail :: Maybe Text

The status message.

targetProductId :: Maybe Text

The identifier of the copied product.

copyProductStatus :: Maybe CopyProductStatus

The status of the copy product operation.

httpStatus :: Int

The response's http status code.

newDescribeCopyProductStatusResponse

Arguments

:: Int
$sel:httpStatus:DescribeCopyProductStatusResponse'

-> DescribeCopyProductStatusResponse

Create a value of DescribeCopyProductStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusDetail:DescribeCopyProductStatusResponse', describeCopyProductStatusResponse_statusDetail - The status message.

$sel:targetProductId:DescribeCopyProductStatusResponse', describeCopyProductStatusResponse_targetProductId - The identifier of the copied product.

$sel:copyProductStatus:DescribeCopyProductStatusResponse', describeCopyProductStatusResponse_copyProductStatus - The status of the copy product operation.

$sel:httpStatus:DescribeCopyProductStatusResponse', describeCopyProductStatusResponse_httpStatus - The response's http status code.

Response Lenses

describeCopyProductStatusResponse_statusDetail :: Lens' DescribeCopyProductStatusResponse (Maybe Text)

The status message.

describeCopyProductStatusResponse_targetProductId :: Lens' DescribeCopyProductStatusResponse (Maybe Text)

The identifier of the copied product.

describeCopyProductStatusResponse_copyProductStatus :: Lens' DescribeCopyProductStatusResponse (Maybe CopyProductStatus)

The status of the copy product operation.

describeCopyProductStatusResponse_httpStatus :: Lens' DescribeCopyProductStatusResponse Int

The response's http status code.