Network.AWS.ServiceCatalog.DescribePortfolioShareStatus

Description

Gets the status of the specified portfolio share operation. This API can only be called by the management account in the organization or by a delegated admin.

Creating a Request

data DescribePortfolioShareStatus

See: newDescribePortfolioShareStatus smart constructor.

Constructors

DescribePortfolioShareStatus'

Fields

portfolioShareToken :: Text

The token for the portfolio share operation. This token is returned either by CreatePortfolioShare or by DeletePortfolioShare.

newDescribePortfolioShareStatus

Arguments

:: Text
$sel:portfolioShareToken:DescribePortfolioShareStatus'

-> DescribePortfolioShareStatus

Create a value of DescribePortfolioShareStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:portfolioShareToken:DescribePortfolioShareStatus', describePortfolioShareStatus_portfolioShareToken - The token for the portfolio share operation. This token is returned either by CreatePortfolioShare or by DeletePortfolioShare.

Request Lenses

describePortfolioShareStatus_portfolioShareToken :: Lens' DescribePortfolioShareStatus Text

The token for the portfolio share operation. This token is returned either by CreatePortfolioShare or by DeletePortfolioShare.

Destructuring the Response

data DescribePortfolioShareStatusResponse

See: newDescribePortfolioShareStatusResponse smart constructor.

Constructors

DescribePortfolioShareStatusResponse'

Fields

shareDetails :: Maybe ShareDetails

Information about the portfolio share operation.

status :: Maybe ShareStatus

Status of the portfolio share operation.

portfolioId :: Maybe Text

The portfolio identifier.

portfolioShareToken :: Maybe Text

The token for the portfolio share operation. For example, share-6v24abcdefghi.

organizationNodeValue :: Maybe Text

Organization node identifier. It can be either account id, organizational unit id or organization id.

httpStatus :: Int

The response's http status code.

newDescribePortfolioShareStatusResponse

Arguments

:: Int
$sel:httpStatus:DescribePortfolioShareStatusResponse'

-> DescribePortfolioShareStatusResponse

Create a value of DescribePortfolioShareStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shareDetails:DescribePortfolioShareStatusResponse', describePortfolioShareStatusResponse_shareDetails - Information about the portfolio share operation.

$sel:status:DescribePortfolioShareStatusResponse', describePortfolioShareStatusResponse_status - Status of the portfolio share operation.

$sel:portfolioId:DescribePortfolioShareStatusResponse', describePortfolioShareStatusResponse_portfolioId - The portfolio identifier.

$sel:portfolioShareToken:DescribePortfolioShareStatus', describePortfolioShareStatusResponse_portfolioShareToken - The token for the portfolio share operation. For example, share-6v24abcdefghi.

$sel:organizationNodeValue:DescribePortfolioShareStatusResponse', describePortfolioShareStatusResponse_organizationNodeValue - Organization node identifier. It can be either account id, organizational unit id or organization id.

$sel:httpStatus:DescribePortfolioShareStatusResponse', describePortfolioShareStatusResponse_httpStatus - The response's http status code.

Response Lenses

describePortfolioShareStatusResponse_shareDetails :: Lens' DescribePortfolioShareStatusResponse (Maybe ShareDetails)

Information about the portfolio share operation.

describePortfolioShareStatusResponse_status :: Lens' DescribePortfolioShareStatusResponse (Maybe ShareStatus)

Status of the portfolio share operation.

describePortfolioShareStatusResponse_portfolioId :: Lens' DescribePortfolioShareStatusResponse (Maybe Text)

The portfolio identifier.

describePortfolioShareStatusResponse_portfolioShareToken :: Lens' DescribePortfolioShareStatusResponse (Maybe Text)

The token for the portfolio share operation. For example, share-6v24abcdefghi.

describePortfolioShareStatusResponse_organizationNodeValue :: Lens' DescribePortfolioShareStatusResponse (Maybe Text)

Organization node identifier. It can be either account id, organizational unit id or organization id.

describePortfolioShareStatusResponse_httpStatus :: Lens' DescribePortfolioShareStatusResponse Int

The response's http status code.