Network.AWS.ServiceCatalog.DescribeProduct

Description

Gets information about the specified product.

Creating a Request

data DescribeProduct

See: newDescribeProduct smart constructor.

Constructors

DescribeProduct'

Fields

id :: Maybe Text

The product identifier.

name :: Maybe Text

The product name.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

newDescribeProduct :: DescribeProduct

Create a value of DescribeProduct with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DescribeProduct', describeProduct_id - The product identifier.

$sel:name:DescribeProduct', describeProduct_name - The product name.

$sel:acceptLanguage:DescribeProduct', describeProduct_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

Request Lenses

describeProduct_id :: Lens' DescribeProduct (Maybe Text)

The product identifier.

describeProduct_name :: Lens' DescribeProduct (Maybe Text)

The product name.

describeProduct_acceptLanguage :: Lens' DescribeProduct (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

Destructuring the Response

data DescribeProductResponse

See: newDescribeProductResponse smart constructor.

Constructors

DescribeProductResponse'

Fields

provisioningArtifacts :: Maybe [ProvisioningArtifact]

Information about the provisioning artifacts for the specified product.

launchPaths :: Maybe [LaunchPath]

Information about the associated launch paths.

productViewSummary :: Maybe ProductViewSummary

Summary information about the product view.

budgets :: Maybe [BudgetDetail]

Information about the associated budgets.

httpStatus :: Int

The response's http status code.

newDescribeProductResponse

Arguments

:: Int
$sel:httpStatus:DescribeProductResponse'

-> DescribeProductResponse

Create a value of DescribeProductResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:provisioningArtifacts:DescribeProductResponse', describeProductResponse_provisioningArtifacts - Information about the provisioning artifacts for the specified product.

$sel:launchPaths:DescribeProductResponse', describeProductResponse_launchPaths - Information about the associated launch paths.

$sel:productViewSummary:DescribeProductResponse', describeProductResponse_productViewSummary - Summary information about the product view.

$sel:budgets:DescribeProductResponse', describeProductResponse_budgets - Information about the associated budgets.

$sel:httpStatus:DescribeProductResponse', describeProductResponse_httpStatus - The response's http status code.

Response Lenses

describeProductResponse_provisioningArtifacts :: Lens' DescribeProductResponse (Maybe [ProvisioningArtifact])

Information about the provisioning artifacts for the specified product.

describeProductResponse_launchPaths :: Lens' DescribeProductResponse (Maybe [LaunchPath])

Information about the associated launch paths.

describeProductResponse_productViewSummary :: Lens' DescribeProductResponse (Maybe ProductViewSummary)

Summary information about the product view.

describeProductResponse_budgets :: Lens' DescribeProductResponse (Maybe [BudgetDetail])

Information about the associated budgets.

describeProductResponse_httpStatus :: Lens' DescribeProductResponse Int

The response's http status code.