Network.AWS.ServiceCatalog.DescribeProductAsAdmin

Description

Gets information about the specified product. This operation is run with administrator access.

Creating a Request

data DescribeProductAsAdmin

See: newDescribeProductAsAdmin smart constructor.

Constructors

DescribeProductAsAdmin'

Fields

id :: Maybe Text

The product identifier.

name :: Maybe Text

The product name.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

sourcePortfolioId :: Maybe Text

The unique identifier of the shared portfolio that the specified product is associated with.

You can provide this parameter to retrieve the shared TagOptions associated with the product. If this parameter is provided and if TagOptions sharing is enabled in the portfolio share, the API returns both local and shared TagOptions associated with the product. Otherwise only local TagOptions will be returned.

newDescribeProductAsAdmin :: DescribeProductAsAdmin

Create a value of DescribeProductAsAdmin with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DescribeProductAsAdmin', describeProductAsAdmin_id - The product identifier.

$sel:name:DescribeProductAsAdmin', describeProductAsAdmin_name - The product name.

$sel:acceptLanguage:DescribeProductAsAdmin', describeProductAsAdmin_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:sourcePortfolioId:DescribeProductAsAdmin', describeProductAsAdmin_sourcePortfolioId - The unique identifier of the shared portfolio that the specified product is associated with.

You can provide this parameter to retrieve the shared TagOptions associated with the product. If this parameter is provided and if TagOptions sharing is enabled in the portfolio share, the API returns both local and shared TagOptions associated with the product. Otherwise only local TagOptions will be returned.

Request Lenses

describeProductAsAdmin_id :: Lens' DescribeProductAsAdmin (Maybe Text)

The product identifier.

describeProductAsAdmin_name :: Lens' DescribeProductAsAdmin (Maybe Text)

The product name.

describeProductAsAdmin_acceptLanguage :: Lens' DescribeProductAsAdmin (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

describeProductAsAdmin_sourcePortfolioId :: Lens' DescribeProductAsAdmin (Maybe Text)

The unique identifier of the shared portfolio that the specified product is associated with.

You can provide this parameter to retrieve the shared TagOptions associated with the product. If this parameter is provided and if TagOptions sharing is enabled in the portfolio share, the API returns both local and shared TagOptions associated with the product. Otherwise only local TagOptions will be returned.

Destructuring the Response

data DescribeProductAsAdminResponse

See: newDescribeProductAsAdminResponse smart constructor.

Constructors

DescribeProductAsAdminResponse'

Fields

productViewDetail :: Maybe ProductViewDetail

Information about the product view.

tags :: Maybe [Tag]

Information about the tags associated with the product.

budgets :: Maybe [BudgetDetail]

Information about the associated budgets.

provisioningArtifactSummaries :: Maybe [ProvisioningArtifactSummary]

Information about the provisioning artifacts (also known as versions) for the specified product.

tagOptions :: Maybe [TagOptionDetail]

Information about the TagOptions associated with the product.

httpStatus :: Int

The response's http status code.

newDescribeProductAsAdminResponse

Arguments

:: Int
$sel:httpStatus:DescribeProductAsAdminResponse'

-> DescribeProductAsAdminResponse

Create a value of DescribeProductAsAdminResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:productViewDetail:DescribeProductAsAdminResponse', describeProductAsAdminResponse_productViewDetail - Information about the product view.

$sel:tags:DescribeProductAsAdminResponse', describeProductAsAdminResponse_tags - Information about the tags associated with the product.

$sel:budgets:DescribeProductAsAdminResponse', describeProductAsAdminResponse_budgets - Information about the associated budgets.

$sel:provisioningArtifactSummaries:DescribeProductAsAdminResponse', describeProductAsAdminResponse_provisioningArtifactSummaries - Information about the provisioning artifacts (also known as versions) for the specified product.

$sel:tagOptions:DescribeProductAsAdminResponse', describeProductAsAdminResponse_tagOptions - Information about the TagOptions associated with the product.

$sel:httpStatus:DescribeProductAsAdminResponse', describeProductAsAdminResponse_httpStatus - The response's http status code.

Response Lenses

describeProductAsAdminResponse_productViewDetail :: Lens' DescribeProductAsAdminResponse (Maybe ProductViewDetail)

Information about the product view.

describeProductAsAdminResponse_tags :: Lens' DescribeProductAsAdminResponse (Maybe [Tag])

Information about the tags associated with the product.

describeProductAsAdminResponse_budgets :: Lens' DescribeProductAsAdminResponse (Maybe [BudgetDetail])

Information about the associated budgets.

describeProductAsAdminResponse_provisioningArtifactSummaries :: Lens' DescribeProductAsAdminResponse (Maybe [ProvisioningArtifactSummary])

Information about the provisioning artifacts (also known as versions) for the specified product.

describeProductAsAdminResponse_tagOptions :: Lens' DescribeProductAsAdminResponse (Maybe [TagOptionDetail])

Information about the TagOptions associated with the product.

describeProductAsAdminResponse_httpStatus :: Lens' DescribeProductAsAdminResponse Int

The response's http status code.