Network.AWS.ServiceCatalog.DescribeProvisionedProductPlan

Description

Gets information about the resource changes for the specified plan.

Creating a Request

data DescribeProvisionedProductPlan

See: newDescribeProvisionedProductPlan smart constructor.

Constructors

DescribeProvisionedProductPlan'

Fields

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

pageToken :: Maybe Text

The page token for the next set of results. To retrieve the first set of results, use null.

pageSize :: Maybe Natural

The maximum number of items to return with this call.

planId :: Text

The plan identifier.

newDescribeProvisionedProductPlan

Arguments

:: Text
$sel:planId:DescribeProvisionedProductPlan'

-> DescribeProvisionedProductPlan

Create a value of DescribeProvisionedProductPlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceptLanguage:DescribeProvisionedProductPlan', describeProvisionedProductPlan_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:pageToken:DescribeProvisionedProductPlan', describeProvisionedProductPlan_pageToken - The page token for the next set of results. To retrieve the first set of results, use null.

$sel:pageSize:DescribeProvisionedProductPlan', describeProvisionedProductPlan_pageSize - The maximum number of items to return with this call.

$sel:planId:DescribeProvisionedProductPlan', describeProvisionedProductPlan_planId - The plan identifier.

Request Lenses

describeProvisionedProductPlan_acceptLanguage :: Lens' DescribeProvisionedProductPlan (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

describeProvisionedProductPlan_pageToken :: Lens' DescribeProvisionedProductPlan (Maybe Text)

The page token for the next set of results. To retrieve the first set of results, use null.

describeProvisionedProductPlan_pageSize :: Lens' DescribeProvisionedProductPlan (Maybe Natural)

The maximum number of items to return with this call.

describeProvisionedProductPlan_planId :: Lens' DescribeProvisionedProductPlan Text

The plan identifier.

Destructuring the Response

data DescribeProvisionedProductPlanResponse

See: newDescribeProvisionedProductPlanResponse smart constructor.

Constructors

DescribeProvisionedProductPlanResponse'

Fields

nextPageToken :: Maybe Text

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

provisionedProductPlanDetails :: Maybe ProvisionedProductPlanDetails

Information about the plan.

resourceChanges :: Maybe [ResourceChange]

Information about the resource changes that will occur when the plan is executed.

httpStatus :: Int

The response's http status code.

newDescribeProvisionedProductPlanResponse

Arguments

:: Int
$sel:httpStatus:DescribeProvisionedProductPlanResponse'

-> DescribeProvisionedProductPlanResponse

Create a value of DescribeProvisionedProductPlanResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:DescribeProvisionedProductPlanResponse', describeProvisionedProductPlanResponse_nextPageToken - The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

$sel:provisionedProductPlanDetails:DescribeProvisionedProductPlanResponse', describeProvisionedProductPlanResponse_provisionedProductPlanDetails - Information about the plan.

$sel:resourceChanges:DescribeProvisionedProductPlanResponse', describeProvisionedProductPlanResponse_resourceChanges - Information about the resource changes that will occur when the plan is executed.

$sel:httpStatus:DescribeProvisionedProductPlanResponse', describeProvisionedProductPlanResponse_httpStatus - The response's http status code.

Response Lenses

describeProvisionedProductPlanResponse_nextPageToken :: Lens' DescribeProvisionedProductPlanResponse (Maybe Text)

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

describeProvisionedProductPlanResponse_provisionedProductPlanDetails :: Lens' DescribeProvisionedProductPlanResponse (Maybe ProvisionedProductPlanDetails)

Information about the plan.

describeProvisionedProductPlanResponse_resourceChanges :: Lens' DescribeProvisionedProductPlanResponse (Maybe [ResourceChange])

Information about the resource changes that will occur when the plan is executed.

describeProvisionedProductPlanResponse_httpStatus :: Lens' DescribeProvisionedProductPlanResponse Int

The response's http status code.