Network.AWS.ServiceCatalog.DescribeProvisioningArtifact

Description

Gets information about the specified provisioning artifact (also known as a version) for the specified product.

Creating a Request

data DescribeProvisioningArtifact

See: newDescribeProvisioningArtifact smart constructor.

Constructors

DescribeProvisioningArtifact'

Fields

provisioningArtifactName :: Maybe Text

The provisioning artifact name.

provisioningArtifactId :: Maybe Text

The identifier of the provisioning artifact.

productName :: Maybe Text

The product name.

productId :: Maybe Text

The product identifier.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

verbose :: Maybe Bool

Indicates whether a verbose level of detail is enabled.

newDescribeProvisioningArtifact :: DescribeProvisioningArtifact

Create a value of DescribeProvisioningArtifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:provisioningArtifactName:DescribeProvisioningArtifact', describeProvisioningArtifact_provisioningArtifactName - The provisioning artifact name.

$sel:provisioningArtifactId:DescribeProvisioningArtifact', describeProvisioningArtifact_provisioningArtifactId - The identifier of the provisioning artifact.

$sel:productName:DescribeProvisioningArtifact', describeProvisioningArtifact_productName - The product name.

$sel:productId:DescribeProvisioningArtifact', describeProvisioningArtifact_productId - The product identifier.

$sel:acceptLanguage:DescribeProvisioningArtifact', describeProvisioningArtifact_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:verbose:DescribeProvisioningArtifact', describeProvisioningArtifact_verbose - Indicates whether a verbose level of detail is enabled.

Request Lenses

describeProvisioningArtifact_provisioningArtifactName :: Lens' DescribeProvisioningArtifact (Maybe Text)

The provisioning artifact name.

describeProvisioningArtifact_provisioningArtifactId :: Lens' DescribeProvisioningArtifact (Maybe Text)

The identifier of the provisioning artifact.

describeProvisioningArtifact_productName :: Lens' DescribeProvisioningArtifact (Maybe Text)

The product name.

describeProvisioningArtifact_productId :: Lens' DescribeProvisioningArtifact (Maybe Text)

The product identifier.

describeProvisioningArtifact_acceptLanguage :: Lens' DescribeProvisioningArtifact (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

describeProvisioningArtifact_verbose :: Lens' DescribeProvisioningArtifact (Maybe Bool)

Indicates whether a verbose level of detail is enabled.

Destructuring the Response

data DescribeProvisioningArtifactResponse

See: newDescribeProvisioningArtifactResponse smart constructor.

Constructors

DescribeProvisioningArtifactResponse'

Fields

status :: Maybe RequestStatus

The status of the current request.

info :: Maybe (HashMap Text Text)

The URL of the CloudFormation template in Amazon S3.

provisioningArtifactDetail :: Maybe ProvisioningArtifactDetail

Information about the provisioning artifact.

httpStatus :: Int

The response's http status code.

newDescribeProvisioningArtifactResponse

Arguments

:: Int
$sel:httpStatus:DescribeProvisioningArtifactResponse'

-> DescribeProvisioningArtifactResponse

Create a value of DescribeProvisioningArtifactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeProvisioningArtifactResponse', describeProvisioningArtifactResponse_status - The status of the current request.

$sel:info:DescribeProvisioningArtifactResponse', describeProvisioningArtifactResponse_info - The URL of the CloudFormation template in Amazon S3.

$sel:provisioningArtifactDetail:DescribeProvisioningArtifactResponse', describeProvisioningArtifactResponse_provisioningArtifactDetail - Information about the provisioning artifact.

$sel:httpStatus:DescribeProvisioningArtifactResponse', describeProvisioningArtifactResponse_httpStatus - The response's http status code.

Response Lenses

describeProvisioningArtifactResponse_status :: Lens' DescribeProvisioningArtifactResponse (Maybe RequestStatus)

The status of the current request.

describeProvisioningArtifactResponse_info :: Lens' DescribeProvisioningArtifactResponse (Maybe (HashMap Text Text))

The URL of the CloudFormation template in Amazon S3.

describeProvisioningArtifactResponse_provisioningArtifactDetail :: Lens' DescribeProvisioningArtifactResponse (Maybe ProvisioningArtifactDetail)

Information about the provisioning artifact.

describeProvisioningArtifactResponse_httpStatus :: Lens' DescribeProvisioningArtifactResponse Int

The response's http status code.