Network.AWS.ServiceCatalog.DescribeProvisioningParameters

Description

Gets information about the configuration required to provision the specified product using the specified provisioning artifact.

If the output contains a TagOption key with an empty list of values, there is a TagOption conflict for that key. The end user cannot take action to fix the conflict, and launch is not blocked. In subsequent calls to ProvisionProduct, do not include conflicted TagOption keys as tags, or this causes the error "Parameter validation failed: Missing required parameter in Tags[N]:Value". Tag the provisioned product with the value sc-tagoption-conflict-portfolioId-productId.

Creating a Request

data DescribeProvisioningParameters

See: newDescribeProvisioningParameters smart constructor.

Constructors

DescribeProvisioningParameters'

Fields

provisioningArtifactName :: Maybe Text

The name of the provisioning artifact. You must provide the name or ID, but not both.

provisioningArtifactId :: Maybe Text

The identifier of the provisioning artifact. You must provide the name or ID, but not both.

productName :: Maybe Text

The name of the product. You must provide the name or ID, but not both.

productId :: Maybe Text

The product identifier. You must provide the product name or ID, but not both.

pathId :: Maybe Text

The path identifier of the product. This value is optional if the product has a default path, and required if the product has more than one path. To list the paths for a product, use ListLaunchPaths. You must provide the name or ID, but not both.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

pathName :: Maybe Text

The name of the path. You must provide the name or ID, but not both.

newDescribeProvisioningParameters :: DescribeProvisioningParameters

Create a value of DescribeProvisioningParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:provisioningArtifactName:DescribeProvisioningParameters', describeProvisioningParameters_provisioningArtifactName - The name of the provisioning artifact. You must provide the name or ID, but not both.

$sel:provisioningArtifactId:DescribeProvisioningParameters', describeProvisioningParameters_provisioningArtifactId - The identifier of the provisioning artifact. You must provide the name or ID, but not both.

$sel:productName:DescribeProvisioningParameters', describeProvisioningParameters_productName - The name of the product. You must provide the name or ID, but not both.

$sel:productId:DescribeProvisioningParameters', describeProvisioningParameters_productId - The product identifier. You must provide the product name or ID, but not both.

$sel:pathId:DescribeProvisioningParameters', describeProvisioningParameters_pathId - The path identifier of the product. This value is optional if the product has a default path, and required if the product has more than one path. To list the paths for a product, use ListLaunchPaths. You must provide the name or ID, but not both.

$sel:acceptLanguage:DescribeProvisioningParameters', describeProvisioningParameters_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:pathName:DescribeProvisioningParameters', describeProvisioningParameters_pathName - The name of the path. You must provide the name or ID, but not both.

Request Lenses

describeProvisioningParameters_provisioningArtifactName :: Lens' DescribeProvisioningParameters (Maybe Text)

The name of the provisioning artifact. You must provide the name or ID, but not both.

describeProvisioningParameters_provisioningArtifactId :: Lens' DescribeProvisioningParameters (Maybe Text)

The identifier of the provisioning artifact. You must provide the name or ID, but not both.

describeProvisioningParameters_productName :: Lens' DescribeProvisioningParameters (Maybe Text)

The name of the product. You must provide the name or ID, but not both.

describeProvisioningParameters_productId :: Lens' DescribeProvisioningParameters (Maybe Text)

The product identifier. You must provide the product name or ID, but not both.

describeProvisioningParameters_pathId :: Lens' DescribeProvisioningParameters (Maybe Text)

The path identifier of the product. This value is optional if the product has a default path, and required if the product has more than one path. To list the paths for a product, use ListLaunchPaths. You must provide the name or ID, but not both.

describeProvisioningParameters_acceptLanguage :: Lens' DescribeProvisioningParameters (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

describeProvisioningParameters_pathName :: Lens' DescribeProvisioningParameters (Maybe Text)

The name of the path. You must provide the name or ID, but not both.

Destructuring the Response

data DescribeProvisioningParametersResponse

See: newDescribeProvisioningParametersResponse smart constructor.

Constructors

DescribeProvisioningParametersResponse'

Fields

constraintSummaries :: Maybe [ConstraintSummary]

Information about the constraints used to provision the product.

usageInstructions :: Maybe [UsageInstruction]

Any additional metadata specifically related to the provisioning of the product. For example, see the Version field of the CloudFormation template.

provisioningArtifactOutputs :: Maybe [ProvisioningArtifactOutput]

The output of the provisioning artifact.

provisioningArtifactPreferences :: Maybe ProvisioningArtifactPreferences

An object that contains information about preferences, such as regions and accounts, for the provisioning artifact.

provisioningArtifactParameters :: Maybe [ProvisioningArtifactParameter]

Information about the parameters used to provision the product.

tagOptions :: Maybe [TagOptionSummary]

Information about the TagOptions associated with the resource.

httpStatus :: Int

The response's http status code.

newDescribeProvisioningParametersResponse

Arguments

:: Int
$sel:httpStatus:DescribeProvisioningParametersResponse'

-> DescribeProvisioningParametersResponse

Create a value of DescribeProvisioningParametersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:constraintSummaries:DescribeProvisioningParametersResponse', describeProvisioningParametersResponse_constraintSummaries - Information about the constraints used to provision the product.

$sel:usageInstructions:DescribeProvisioningParametersResponse', describeProvisioningParametersResponse_usageInstructions - Any additional metadata specifically related to the provisioning of the product. For example, see the Version field of the CloudFormation template.

$sel:provisioningArtifactOutputs:DescribeProvisioningParametersResponse', describeProvisioningParametersResponse_provisioningArtifactOutputs - The output of the provisioning artifact.

$sel:provisioningArtifactPreferences:DescribeProvisioningParametersResponse', describeProvisioningParametersResponse_provisioningArtifactPreferences - An object that contains information about preferences, such as regions and accounts, for the provisioning artifact.

$sel:provisioningArtifactParameters:DescribeProvisioningParametersResponse', describeProvisioningParametersResponse_provisioningArtifactParameters - Information about the parameters used to provision the product.

$sel:tagOptions:DescribeProvisioningParametersResponse', describeProvisioningParametersResponse_tagOptions - Information about the TagOptions associated with the resource.

$sel:httpStatus:DescribeProvisioningParametersResponse', describeProvisioningParametersResponse_httpStatus - The response's http status code.

Response Lenses

describeProvisioningParametersResponse_constraintSummaries :: Lens' DescribeProvisioningParametersResponse (Maybe [ConstraintSummary])

Information about the constraints used to provision the product.

describeProvisioningParametersResponse_usageInstructions :: Lens' DescribeProvisioningParametersResponse (Maybe [UsageInstruction])

Any additional metadata specifically related to the provisioning of the product. For example, see the Version field of the CloudFormation template.

describeProvisioningParametersResponse_provisioningArtifactOutputs :: Lens' DescribeProvisioningParametersResponse (Maybe [ProvisioningArtifactOutput])

The output of the provisioning artifact.

describeProvisioningParametersResponse_provisioningArtifactPreferences :: Lens' DescribeProvisioningParametersResponse (Maybe ProvisioningArtifactPreferences)

An object that contains information about preferences, such as regions and accounts, for the provisioning artifact.

describeProvisioningParametersResponse_provisioningArtifactParameters :: Lens' DescribeProvisioningParametersResponse (Maybe [ProvisioningArtifactParameter])

Information about the parameters used to provision the product.

describeProvisioningParametersResponse_tagOptions :: Lens' DescribeProvisioningParametersResponse (Maybe [TagOptionSummary])

Information about the TagOptions associated with the resource.

describeProvisioningParametersResponse_httpStatus :: Lens' DescribeProvisioningParametersResponse Int

The response's http status code.