Network.AWS.ServiceCatalog.DisableAWSOrganizationsAccess

Description

Disable portfolio sharing through AWS Organizations feature. This feature will not delete your current shares but it will prevent you from creating new shares throughout your organization. Current shares will not be in sync with your organization structure if it changes after calling this API. This API can only be called by the management account in the organization.

This API can't be invoked if there are active delegated administrators in the organization.

Note that a delegated administrator is not authorized to invoke DisableAWSOrganizationsAccess.

Creating a Request

data DisableAWSOrganizationsAccess

See: newDisableAWSOrganizationsAccess smart constructor.

Constructors

DisableAWSOrganizationsAccess'

newDisableAWSOrganizationsAccess :: DisableAWSOrganizationsAccess

Create a value of DisableAWSOrganizationsAccess with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data DisableAWSOrganizationsAccessResponse

See: newDisableAWSOrganizationsAccessResponse smart constructor.

Constructors

DisableAWSOrganizationsAccessResponse'

Fields

httpStatus :: Int

The response's http status code.

newDisableAWSOrganizationsAccessResponse

Arguments

:: Int
$sel:httpStatus:DisableAWSOrganizationsAccessResponse'

-> DisableAWSOrganizationsAccessResponse

Create a value of DisableAWSOrganizationsAccessResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisableAWSOrganizationsAccessResponse', disableAWSOrganizationsAccessResponse_httpStatus - The response's http status code.

Response Lenses

disableAWSOrganizationsAccessResponse_httpStatus :: Lens' DisableAWSOrganizationsAccessResponse Int

The response's http status code.