Network.AWS.ServiceCatalog.DisassociatePrincipalFromPortfolio

Description

Disassociates a previously associated principal ARN from a specified portfolio.

Creating a Request

data DisassociatePrincipalFromPortfolio

See: newDisassociatePrincipalFromPortfolio smart constructor.

Constructors

DisassociatePrincipalFromPortfolio'

Fields

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

portfolioId :: Text

The portfolio identifier.

principalARN :: Text

The ARN of the principal (IAM user, role, or group).

newDisassociatePrincipalFromPortfolio

Arguments

:: Text
$sel:portfolioId:DisassociatePrincipalFromPortfolio'

-> Text
$sel:principalARN:DisassociatePrincipalFromPortfolio'

-> DisassociatePrincipalFromPortfolio

Create a value of DisassociatePrincipalFromPortfolio with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceptLanguage:DisassociatePrincipalFromPortfolio', disassociatePrincipalFromPortfolio_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:portfolioId:DisassociatePrincipalFromPortfolio', disassociatePrincipalFromPortfolio_portfolioId - The portfolio identifier.

$sel:principalARN:DisassociatePrincipalFromPortfolio', disassociatePrincipalFromPortfolio_principalARN - The ARN of the principal (IAM user, role, or group).

Request Lenses

disassociatePrincipalFromPortfolio_acceptLanguage :: Lens' DisassociatePrincipalFromPortfolio (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

disassociatePrincipalFromPortfolio_portfolioId :: Lens' DisassociatePrincipalFromPortfolio Text

The portfolio identifier.

disassociatePrincipalFromPortfolio_principalARN :: Lens' DisassociatePrincipalFromPortfolio Text

The ARN of the principal (IAM user, role, or group).

Destructuring the Response

data DisassociatePrincipalFromPortfolioResponse

See: newDisassociatePrincipalFromPortfolioResponse smart constructor.

Constructors

DisassociatePrincipalFromPortfolioResponse'

Fields

httpStatus :: Int

The response's http status code.

newDisassociatePrincipalFromPortfolioResponse

Arguments

:: Int
$sel:httpStatus:DisassociatePrincipalFromPortfolioResponse'

-> DisassociatePrincipalFromPortfolioResponse

Create a value of DisassociatePrincipalFromPortfolioResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisassociatePrincipalFromPortfolioResponse', disassociatePrincipalFromPortfolioResponse_httpStatus - The response's http status code.

Response Lenses

disassociatePrincipalFromPortfolioResponse_httpStatus :: Lens' DisassociatePrincipalFromPortfolioResponse Int

The response's http status code.