Network.AWS.ServiceCatalog.DisassociateProductFromPortfolio

Description

Disassociates the specified product from the specified portfolio.

A delegated admin is authorized to invoke this command.

Creating a Request

data DisassociateProductFromPortfolio

See: newDisassociateProductFromPortfolio smart constructor.

Constructors

DisassociateProductFromPortfolio'

Fields

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

productId :: Text

The product identifier.

portfolioId :: Text

The portfolio identifier.

newDisassociateProductFromPortfolio

Arguments

:: Text
$sel:productId:DisassociateProductFromPortfolio'

-> Text
$sel:portfolioId:DisassociateProductFromPortfolio'

-> DisassociateProductFromPortfolio

Create a value of DisassociateProductFromPortfolio with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceptLanguage:DisassociateProductFromPortfolio', disassociateProductFromPortfolio_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:productId:DisassociateProductFromPortfolio', disassociateProductFromPortfolio_productId - The product identifier.

$sel:portfolioId:DisassociateProductFromPortfolio', disassociateProductFromPortfolio_portfolioId - The portfolio identifier.

Request Lenses

disassociateProductFromPortfolio_acceptLanguage :: Lens' DisassociateProductFromPortfolio (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

disassociateProductFromPortfolio_productId :: Lens' DisassociateProductFromPortfolio Text

The product identifier.

disassociateProductFromPortfolio_portfolioId :: Lens' DisassociateProductFromPortfolio Text

The portfolio identifier.

Destructuring the Response

data DisassociateProductFromPortfolioResponse

See: newDisassociateProductFromPortfolioResponse smart constructor.

Constructors

DisassociateProductFromPortfolioResponse'

Fields

httpStatus :: Int

The response's http status code.

newDisassociateProductFromPortfolioResponse

Arguments

:: Int
$sel:httpStatus:DisassociateProductFromPortfolioResponse'

-> DisassociateProductFromPortfolioResponse

Create a value of DisassociateProductFromPortfolioResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisassociateProductFromPortfolioResponse', disassociateProductFromPortfolioResponse_httpStatus - The response's http status code.

Response Lenses

disassociateProductFromPortfolioResponse_httpStatus :: Lens' DisassociateProductFromPortfolioResponse Int

The response's http status code.