Network.AWS.ServiceCatalog.DisassociateServiceActionFromProvisioningArtifact

Description

Disassociates the specified self-service action association from the specified provisioning artifact.

Creating a Request

data DisassociateServiceActionFromProvisioningArtifact

See: newDisassociateServiceActionFromProvisioningArtifact smart constructor.

Constructors

DisassociateServiceActionFromProvisioningArtifact'

Fields

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

productId :: Text

The product identifier. For example, prod-abcdzk7xy33qa.

provisioningArtifactId :: Text

The identifier of the provisioning artifact. For example, pa-4abcdjnxjj6ne.

serviceActionId :: Text

The self-service action identifier. For example, act-fs7abcd89wxyz.

newDisassociateServiceActionFromProvisioningArtifact

Arguments

:: Text
$sel:productId:DisassociateServiceActionFromProvisioningArtifact'

-> Text
$sel:provisioningArtifactId:DisassociateServiceActionFromProvisioningArtifact'

-> Text
$sel:serviceActionId:DisassociateServiceActionFromProvisioningArtifact'

-> DisassociateServiceActionFromProvisioningArtifact

Create a value of DisassociateServiceActionFromProvisioningArtifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceptLanguage:DisassociateServiceActionFromProvisioningArtifact', disassociateServiceActionFromProvisioningArtifact_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:productId:DisassociateServiceActionFromProvisioningArtifact', disassociateServiceActionFromProvisioningArtifact_productId - The product identifier. For example, prod-abcdzk7xy33qa.

$sel:provisioningArtifactId:DisassociateServiceActionFromProvisioningArtifact', disassociateServiceActionFromProvisioningArtifact_provisioningArtifactId - The identifier of the provisioning artifact. For example, pa-4abcdjnxjj6ne.

$sel:serviceActionId:DisassociateServiceActionFromProvisioningArtifact', disassociateServiceActionFromProvisioningArtifact_serviceActionId - The self-service action identifier. For example, act-fs7abcd89wxyz.

Request Lenses

disassociateServiceActionFromProvisioningArtifact_acceptLanguage :: Lens' DisassociateServiceActionFromProvisioningArtifact (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

disassociateServiceActionFromProvisioningArtifact_productId :: Lens' DisassociateServiceActionFromProvisioningArtifact Text

The product identifier. For example, prod-abcdzk7xy33qa.

disassociateServiceActionFromProvisioningArtifact_provisioningArtifactId :: Lens' DisassociateServiceActionFromProvisioningArtifact Text

The identifier of the provisioning artifact. For example, pa-4abcdjnxjj6ne.

disassociateServiceActionFromProvisioningArtifact_serviceActionId :: Lens' DisassociateServiceActionFromProvisioningArtifact Text

The self-service action identifier. For example, act-fs7abcd89wxyz.

Destructuring the Response

data DisassociateServiceActionFromProvisioningArtifactResponse

See: newDisassociateServiceActionFromProvisioningArtifactResponse smart constructor.

Constructors

DisassociateServiceActionFromProvisioningArtifactResponse'

Fields

httpStatus :: Int

The response's http status code.

newDisassociateServiceActionFromProvisioningArtifactResponse

Arguments

:: Int
$sel:httpStatus:DisassociateServiceActionFromProvisioningArtifactResponse'

-> DisassociateServiceActionFromProvisioningArtifactResponse

Create a value of DisassociateServiceActionFromProvisioningArtifactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisassociateServiceActionFromProvisioningArtifactResponse', disassociateServiceActionFromProvisioningArtifactResponse_httpStatus - The response's http status code.

Response Lenses

disassociateServiceActionFromProvisioningArtifactResponse_httpStatus :: Lens' DisassociateServiceActionFromProvisioningArtifactResponse Int

The response's http status code.