Network.AWS.ServiceCatalog.ExecuteProvisionedProductServiceAction

Description

Executes a self-service action against a provisioned product.

Creating a Request

data ExecuteProvisionedProductServiceAction

See: newExecuteProvisionedProductServiceAction smart constructor.

Constructors

ExecuteProvisionedProductServiceAction'

Fields

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

parameters :: Maybe (HashMap Text [Text])

A map of all self-service action parameters and their values. If a provided parameter is of a special type, such as TARGET, the provided value will override the default value generated by AWS Service Catalog. If the parameters field is not provided, no additional parameters are passed and default values will be used for any special parameters such as TARGET.

provisionedProductId :: Text

The identifier of the provisioned product.

serviceActionId :: Text

The self-service action identifier. For example, act-fs7abcd89wxyz.

executeToken :: Text

An idempotency token that uniquely identifies the execute request.

newExecuteProvisionedProductServiceAction

Arguments

:: Text
$sel:provisionedProductId:ExecuteProvisionedProductServiceAction'

-> Text
$sel:serviceActionId:ExecuteProvisionedProductServiceAction'

-> Text
$sel:executeToken:ExecuteProvisionedProductServiceAction'

-> ExecuteProvisionedProductServiceAction

Create a value of ExecuteProvisionedProductServiceAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceptLanguage:ExecuteProvisionedProductServiceAction', executeProvisionedProductServiceAction_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:parameters:ExecuteProvisionedProductServiceAction', executeProvisionedProductServiceAction_parameters - A map of all self-service action parameters and their values. If a provided parameter is of a special type, such as TARGET, the provided value will override the default value generated by AWS Service Catalog. If the parameters field is not provided, no additional parameters are passed and default values will be used for any special parameters such as TARGET.

$sel:provisionedProductId:ExecuteProvisionedProductServiceAction', executeProvisionedProductServiceAction_provisionedProductId - The identifier of the provisioned product.

$sel:serviceActionId:ExecuteProvisionedProductServiceAction', executeProvisionedProductServiceAction_serviceActionId - The self-service action identifier. For example, act-fs7abcd89wxyz.

$sel:executeToken:ExecuteProvisionedProductServiceAction', executeProvisionedProductServiceAction_executeToken - An idempotency token that uniquely identifies the execute request.

Request Lenses

executeProvisionedProductServiceAction_acceptLanguage :: Lens' ExecuteProvisionedProductServiceAction (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

executeProvisionedProductServiceAction_parameters :: Lens' ExecuteProvisionedProductServiceAction (Maybe (HashMap Text [Text]))

A map of all self-service action parameters and their values. If a provided parameter is of a special type, such as TARGET, the provided value will override the default value generated by AWS Service Catalog. If the parameters field is not provided, no additional parameters are passed and default values will be used for any special parameters such as TARGET.

executeProvisionedProductServiceAction_provisionedProductId :: Lens' ExecuteProvisionedProductServiceAction Text

The identifier of the provisioned product.

executeProvisionedProductServiceAction_serviceActionId :: Lens' ExecuteProvisionedProductServiceAction Text

The self-service action identifier. For example, act-fs7abcd89wxyz.

executeProvisionedProductServiceAction_executeToken :: Lens' ExecuteProvisionedProductServiceAction Text

An idempotency token that uniquely identifies the execute request.

Destructuring the Response

data ExecuteProvisionedProductServiceActionResponse

See: newExecuteProvisionedProductServiceActionResponse smart constructor.

Constructors

ExecuteProvisionedProductServiceActionResponse'

Fields

recordDetail :: Maybe RecordDetail

An object containing detailed information about the result of provisioning the product.

httpStatus :: Int

The response's http status code.

newExecuteProvisionedProductServiceActionResponse

Arguments

:: Int
$sel:httpStatus:ExecuteProvisionedProductServiceActionResponse'

-> ExecuteProvisionedProductServiceActionResponse

Create a value of ExecuteProvisionedProductServiceActionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recordDetail:ExecuteProvisionedProductServiceActionResponse', executeProvisionedProductServiceActionResponse_recordDetail - An object containing detailed information about the result of provisioning the product.

$sel:httpStatus:ExecuteProvisionedProductServiceActionResponse', executeProvisionedProductServiceActionResponse_httpStatus - The response's http status code.

Response Lenses

executeProvisionedProductServiceActionResponse_recordDetail :: Lens' ExecuteProvisionedProductServiceActionResponse (Maybe RecordDetail)

An object containing detailed information about the result of provisioning the product.

executeProvisionedProductServiceActionResponse_httpStatus :: Lens' ExecuteProvisionedProductServiceActionResponse Int

The response's http status code.