Network.AWS.ServiceCatalog.GetProvisionedProductOutputs

Description

This API takes either a ProvisonedProductId or a ProvisionedProductName, along with a list of one or more output keys, and responds with the key/value pairs of those outputs.

Creating a Request

data GetProvisionedProductOutputs

See: newGetProvisionedProductOutputs smart constructor.

Constructors

GetProvisionedProductOutputs'

Fields

provisionedProductName :: Maybe Text

The name of the provisioned product that you want the outputs from.

pageSize :: Maybe Natural

The maximum number of items to return with this call.

provisionedProductId :: Maybe Text

The identifier of the provisioned product that you want the outputs from.

outputKeys :: Maybe [Text]

The list of keys that the API should return with their values. If none are provided, the API will return all outputs of the provisioned product.

pageToken :: Maybe Text

The page token for the next set of results. To retrieve the first set of results, use null.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

newGetProvisionedProductOutputs :: GetProvisionedProductOutputs

Create a value of GetProvisionedProductOutputs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:provisionedProductName:GetProvisionedProductOutputs', getProvisionedProductOutputs_provisionedProductName - The name of the provisioned product that you want the outputs from.

$sel:pageSize:GetProvisionedProductOutputs', getProvisionedProductOutputs_pageSize - The maximum number of items to return with this call.

$sel:provisionedProductId:GetProvisionedProductOutputs', getProvisionedProductOutputs_provisionedProductId - The identifier of the provisioned product that you want the outputs from.

$sel:outputKeys:GetProvisionedProductOutputs', getProvisionedProductOutputs_outputKeys - The list of keys that the API should return with their values. If none are provided, the API will return all outputs of the provisioned product.

$sel:pageToken:GetProvisionedProductOutputs', getProvisionedProductOutputs_pageToken - The page token for the next set of results. To retrieve the first set of results, use null.

$sel:acceptLanguage:GetProvisionedProductOutputs', getProvisionedProductOutputs_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

Request Lenses

getProvisionedProductOutputs_provisionedProductName :: Lens' GetProvisionedProductOutputs (Maybe Text)

The name of the provisioned product that you want the outputs from.

getProvisionedProductOutputs_pageSize :: Lens' GetProvisionedProductOutputs (Maybe Natural)

The maximum number of items to return with this call.

getProvisionedProductOutputs_provisionedProductId :: Lens' GetProvisionedProductOutputs (Maybe Text)

The identifier of the provisioned product that you want the outputs from.

getProvisionedProductOutputs_outputKeys :: Lens' GetProvisionedProductOutputs (Maybe [Text])

The list of keys that the API should return with their values. If none are provided, the API will return all outputs of the provisioned product.

getProvisionedProductOutputs_pageToken :: Lens' GetProvisionedProductOutputs (Maybe Text)

The page token for the next set of results. To retrieve the first set of results, use null.

getProvisionedProductOutputs_acceptLanguage :: Lens' GetProvisionedProductOutputs (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

Destructuring the Response

data GetProvisionedProductOutputsResponse

See: newGetProvisionedProductOutputsResponse smart constructor.

Constructors

GetProvisionedProductOutputsResponse'

Fields

outputs :: Maybe [RecordOutput]

Information about the product created as the result of a request. For example, the output for a CloudFormation-backed product that creates an S3 bucket would include the S3 bucket URL.

nextPageToken :: Maybe Text

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

httpStatus :: Int

The response's http status code.

newGetProvisionedProductOutputsResponse

Arguments

:: Int
$sel:httpStatus:GetProvisionedProductOutputsResponse'

-> GetProvisionedProductOutputsResponse

Create a value of GetProvisionedProductOutputsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outputs:GetProvisionedProductOutputsResponse', getProvisionedProductOutputsResponse_outputs - Information about the product created as the result of a request. For example, the output for a CloudFormation-backed product that creates an S3 bucket would include the S3 bucket URL.

$sel:nextPageToken:GetProvisionedProductOutputsResponse', getProvisionedProductOutputsResponse_nextPageToken - The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

$sel:httpStatus:GetProvisionedProductOutputsResponse', getProvisionedProductOutputsResponse_httpStatus - The response's http status code.

Response Lenses

getProvisionedProductOutputsResponse_outputs :: Lens' GetProvisionedProductOutputsResponse (Maybe [RecordOutput])

Information about the product created as the result of a request. For example, the output for a CloudFormation-backed product that creates an S3 bucket would include the S3 bucket URL.

getProvisionedProductOutputsResponse_nextPageToken :: Lens' GetProvisionedProductOutputsResponse (Maybe Text)

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

getProvisionedProductOutputsResponse_httpStatus :: Lens' GetProvisionedProductOutputsResponse Int

The response's http status code.