Network.AWS.ServiceCatalog.ListAcceptedPortfolioShares

Description

Lists all portfolios for which sharing was accepted by this account.

This operation returns paginated results.

Creating a Request

data ListAcceptedPortfolioShares

See: newListAcceptedPortfolioShares smart constructor.

Constructors

ListAcceptedPortfolioShares'

Fields

portfolioShareType :: Maybe PortfolioShareType

The type of shared portfolios to list. The default is to list imported portfolios.

AWS_ORGANIZATIONS - List portfolios shared by the management account of your organization

AWS_SERVICECATALOG - List default portfolios

IMPORTED - List imported portfolios

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

pageToken :: Maybe Text

The page token for the next set of results. To retrieve the first set of results, use null.

pageSize :: Maybe Natural

The maximum number of items to return with this call.

newListAcceptedPortfolioShares :: ListAcceptedPortfolioShares

Create a value of ListAcceptedPortfolioShares with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:portfolioShareType:ListAcceptedPortfolioShares', listAcceptedPortfolioShares_portfolioShareType - The type of shared portfolios to list. The default is to list imported portfolios.

AWS_ORGANIZATIONS - List portfolios shared by the management account of your organization

AWS_SERVICECATALOG - List default portfolios

IMPORTED - List imported portfolios

$sel:acceptLanguage:ListAcceptedPortfolioShares', listAcceptedPortfolioShares_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:pageToken:ListAcceptedPortfolioShares', listAcceptedPortfolioShares_pageToken - The page token for the next set of results. To retrieve the first set of results, use null.

$sel:pageSize:ListAcceptedPortfolioShares', listAcceptedPortfolioShares_pageSize - The maximum number of items to return with this call.

Request Lenses

listAcceptedPortfolioShares_portfolioShareType :: Lens' ListAcceptedPortfolioShares (Maybe PortfolioShareType)

The type of shared portfolios to list. The default is to list imported portfolios.

AWS_ORGANIZATIONS - List portfolios shared by the management account of your organization

AWS_SERVICECATALOG - List default portfolios

IMPORTED - List imported portfolios

listAcceptedPortfolioShares_acceptLanguage :: Lens' ListAcceptedPortfolioShares (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

listAcceptedPortfolioShares_pageToken :: Lens' ListAcceptedPortfolioShares (Maybe Text)

The page token for the next set of results. To retrieve the first set of results, use null.

listAcceptedPortfolioShares_pageSize :: Lens' ListAcceptedPortfolioShares (Maybe Natural)

The maximum number of items to return with this call.

Destructuring the Response

data ListAcceptedPortfolioSharesResponse

See: newListAcceptedPortfolioSharesResponse smart constructor.

Constructors

ListAcceptedPortfolioSharesResponse'

Fields

nextPageToken :: Maybe Text

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

portfolioDetails :: Maybe [PortfolioDetail]

Information about the portfolios.

httpStatus :: Int

The response's http status code.

newListAcceptedPortfolioSharesResponse

Arguments

:: Int
$sel:httpStatus:ListAcceptedPortfolioSharesResponse'

-> ListAcceptedPortfolioSharesResponse

Create a value of ListAcceptedPortfolioSharesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:ListAcceptedPortfolioSharesResponse', listAcceptedPortfolioSharesResponse_nextPageToken - The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

$sel:portfolioDetails:ListAcceptedPortfolioSharesResponse', listAcceptedPortfolioSharesResponse_portfolioDetails - Information about the portfolios.

$sel:httpStatus:ListAcceptedPortfolioSharesResponse', listAcceptedPortfolioSharesResponse_httpStatus - The response's http status code.

Response Lenses

listAcceptedPortfolioSharesResponse_nextPageToken :: Lens' ListAcceptedPortfolioSharesResponse (Maybe Text)

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

listAcceptedPortfolioSharesResponse_portfolioDetails :: Lens' ListAcceptedPortfolioSharesResponse (Maybe [PortfolioDetail])

Information about the portfolios.

listAcceptedPortfolioSharesResponse_httpStatus :: Lens' ListAcceptedPortfolioSharesResponse Int

The response's http status code.