Network.AWS.ServiceCatalog.ListConstraintsForPortfolio

Description

Lists the constraints for the specified portfolio and product.

This operation returns paginated results.

Creating a Request

data ListConstraintsForPortfolio

See: newListConstraintsForPortfolio smart constructor.

Constructors

ListConstraintsForPortfolio'

Fields

pageSize :: Maybe Natural

The maximum number of items to return with this call.

pageToken :: Maybe Text

The page token for the next set of results. To retrieve the first set of results, use null.

productId :: Maybe Text

The product identifier.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

portfolioId :: Text

The portfolio identifier.

newListConstraintsForPortfolio

Arguments

:: Text
$sel:portfolioId:ListConstraintsForPortfolio'

-> ListConstraintsForPortfolio

Create a value of ListConstraintsForPortfolio with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageSize:ListConstraintsForPortfolio', listConstraintsForPortfolio_pageSize - The maximum number of items to return with this call.

$sel:pageToken:ListConstraintsForPortfolio', listConstraintsForPortfolio_pageToken - The page token for the next set of results. To retrieve the first set of results, use null.

$sel:productId:ListConstraintsForPortfolio', listConstraintsForPortfolio_productId - The product identifier.

$sel:acceptLanguage:ListConstraintsForPortfolio', listConstraintsForPortfolio_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:portfolioId:ListConstraintsForPortfolio', listConstraintsForPortfolio_portfolioId - The portfolio identifier.

Request Lenses

listConstraintsForPortfolio_pageSize :: Lens' ListConstraintsForPortfolio (Maybe Natural)

The maximum number of items to return with this call.

listConstraintsForPortfolio_pageToken :: Lens' ListConstraintsForPortfolio (Maybe Text)

The page token for the next set of results. To retrieve the first set of results, use null.

listConstraintsForPortfolio_productId :: Lens' ListConstraintsForPortfolio (Maybe Text)

The product identifier.

listConstraintsForPortfolio_acceptLanguage :: Lens' ListConstraintsForPortfolio (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

listConstraintsForPortfolio_portfolioId :: Lens' ListConstraintsForPortfolio Text

The portfolio identifier.

Destructuring the Response

data ListConstraintsForPortfolioResponse

See: newListConstraintsForPortfolioResponse smart constructor.

Constructors

ListConstraintsForPortfolioResponse'

Fields

constraintDetails :: Maybe [ConstraintDetail]

Information about the constraints.

nextPageToken :: Maybe Text

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

httpStatus :: Int

The response's http status code.

newListConstraintsForPortfolioResponse

Arguments

:: Int
$sel:httpStatus:ListConstraintsForPortfolioResponse'

-> ListConstraintsForPortfolioResponse

Create a value of ListConstraintsForPortfolioResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:constraintDetails:ListConstraintsForPortfolioResponse', listConstraintsForPortfolioResponse_constraintDetails - Information about the constraints.

$sel:nextPageToken:ListConstraintsForPortfolioResponse', listConstraintsForPortfolioResponse_nextPageToken - The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

$sel:httpStatus:ListConstraintsForPortfolioResponse', listConstraintsForPortfolioResponse_httpStatus - The response's http status code.

Response Lenses

listConstraintsForPortfolioResponse_constraintDetails :: Lens' ListConstraintsForPortfolioResponse (Maybe [ConstraintDetail])

Information about the constraints.

listConstraintsForPortfolioResponse_nextPageToken :: Lens' ListConstraintsForPortfolioResponse (Maybe Text)

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

listConstraintsForPortfolioResponse_httpStatus :: Lens' ListConstraintsForPortfolioResponse Int

The response's http status code.