Network.AWS.ServiceCatalog.ListPrincipalsForPortfolio

Description

Lists all principal ARNs associated with the specified portfolio.

This operation returns paginated results.

Creating a Request

data ListPrincipalsForPortfolio

See: newListPrincipalsForPortfolio smart constructor.

Constructors

ListPrincipalsForPortfolio'

Fields

pageSize :: Maybe Natural

The maximum number of items to return with this call.

pageToken :: Maybe Text

The page token for the next set of results. To retrieve the first set of results, use null.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

portfolioId :: Text

The portfolio identifier.

newListPrincipalsForPortfolio

Arguments

:: Text
$sel:portfolioId:ListPrincipalsForPortfolio'

-> ListPrincipalsForPortfolio

Create a value of ListPrincipalsForPortfolio with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageSize:ListPrincipalsForPortfolio', listPrincipalsForPortfolio_pageSize - The maximum number of items to return with this call.

$sel:pageToken:ListPrincipalsForPortfolio', listPrincipalsForPortfolio_pageToken - The page token for the next set of results. To retrieve the first set of results, use null.

$sel:acceptLanguage:ListPrincipalsForPortfolio', listPrincipalsForPortfolio_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:portfolioId:ListPrincipalsForPortfolio', listPrincipalsForPortfolio_portfolioId - The portfolio identifier.

Request Lenses

listPrincipalsForPortfolio_pageSize :: Lens' ListPrincipalsForPortfolio (Maybe Natural)

The maximum number of items to return with this call.

listPrincipalsForPortfolio_pageToken :: Lens' ListPrincipalsForPortfolio (Maybe Text)

The page token for the next set of results. To retrieve the first set of results, use null.

listPrincipalsForPortfolio_acceptLanguage :: Lens' ListPrincipalsForPortfolio (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

listPrincipalsForPortfolio_portfolioId :: Lens' ListPrincipalsForPortfolio Text

The portfolio identifier.

Destructuring the Response

data ListPrincipalsForPortfolioResponse

See: newListPrincipalsForPortfolioResponse smart constructor.

Constructors

ListPrincipalsForPortfolioResponse'

Fields

principals :: Maybe [Principal]

The IAM principals (users or roles) associated with the portfolio.

nextPageToken :: Maybe Text

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

httpStatus :: Int

The response's http status code.

newListPrincipalsForPortfolioResponse

Arguments

:: Int
$sel:httpStatus:ListPrincipalsForPortfolioResponse'

-> ListPrincipalsForPortfolioResponse

Create a value of ListPrincipalsForPortfolioResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:principals:ListPrincipalsForPortfolioResponse', listPrincipalsForPortfolioResponse_principals - The IAM principals (users or roles) associated with the portfolio.

$sel:nextPageToken:ListPrincipalsForPortfolioResponse', listPrincipalsForPortfolioResponse_nextPageToken - The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

$sel:httpStatus:ListPrincipalsForPortfolioResponse', listPrincipalsForPortfolioResponse_httpStatus - The response's http status code.

Response Lenses

listPrincipalsForPortfolioResponse_principals :: Lens' ListPrincipalsForPortfolioResponse (Maybe [Principal])

The IAM principals (users or roles) associated with the portfolio.

listPrincipalsForPortfolioResponse_nextPageToken :: Lens' ListPrincipalsForPortfolioResponse (Maybe Text)

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

listPrincipalsForPortfolioResponse_httpStatus :: Lens' ListPrincipalsForPortfolioResponse Int

The response's http status code.