Network.AWS.ServiceCatalog.ListProvisionedProductPlans

Description

Lists the plans for the specified provisioned product or all plans to which the user has access.

This operation returns paginated results.

Creating a Request

data ListProvisionedProductPlans

See: newListProvisionedProductPlans smart constructor.

Constructors

ListProvisionedProductPlans'

Fields

provisionProductId :: Maybe Text

The product identifier.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

accessLevelFilter :: Maybe AccessLevelFilter

The access level to use to obtain results. The default is User.

pageToken :: Maybe Text

The page token for the next set of results. To retrieve the first set of results, use null.

pageSize :: Maybe Natural

The maximum number of items to return with this call.

newListProvisionedProductPlans :: ListProvisionedProductPlans

Create a value of ListProvisionedProductPlans with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:provisionProductId:ListProvisionedProductPlans', listProvisionedProductPlans_provisionProductId - The product identifier.

$sel:acceptLanguage:ListProvisionedProductPlans', listProvisionedProductPlans_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:accessLevelFilter:ListProvisionedProductPlans', listProvisionedProductPlans_accessLevelFilter - The access level to use to obtain results. The default is User.

$sel:pageToken:ListProvisionedProductPlans', listProvisionedProductPlans_pageToken - The page token for the next set of results. To retrieve the first set of results, use null.

$sel:pageSize:ListProvisionedProductPlans', listProvisionedProductPlans_pageSize - The maximum number of items to return with this call.

Request Lenses

listProvisionedProductPlans_provisionProductId :: Lens' ListProvisionedProductPlans (Maybe Text)

The product identifier.

listProvisionedProductPlans_acceptLanguage :: Lens' ListProvisionedProductPlans (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

listProvisionedProductPlans_accessLevelFilter :: Lens' ListProvisionedProductPlans (Maybe AccessLevelFilter)

The access level to use to obtain results. The default is User.

listProvisionedProductPlans_pageToken :: Lens' ListProvisionedProductPlans (Maybe Text)

The page token for the next set of results. To retrieve the first set of results, use null.

listProvisionedProductPlans_pageSize :: Lens' ListProvisionedProductPlans (Maybe Natural)

The maximum number of items to return with this call.

Destructuring the Response

data ListProvisionedProductPlansResponse

See: newListProvisionedProductPlansResponse smart constructor.

Constructors

ListProvisionedProductPlansResponse'

Fields

nextPageToken :: Maybe Text

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

provisionedProductPlans :: Maybe [ProvisionedProductPlanSummary]

Information about the plans.

httpStatus :: Int

The response's http status code.

newListProvisionedProductPlansResponse

Arguments

:: Int
$sel:httpStatus:ListProvisionedProductPlansResponse'

-> ListProvisionedProductPlansResponse

Create a value of ListProvisionedProductPlansResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:ListProvisionedProductPlansResponse', listProvisionedProductPlansResponse_nextPageToken - The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

$sel:provisionedProductPlans:ListProvisionedProductPlansResponse', listProvisionedProductPlansResponse_provisionedProductPlans - Information about the plans.

$sel:httpStatus:ListProvisionedProductPlansResponse', listProvisionedProductPlansResponse_httpStatus - The response's http status code.

Response Lenses

listProvisionedProductPlansResponse_nextPageToken :: Lens' ListProvisionedProductPlansResponse (Maybe Text)

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

listProvisionedProductPlansResponse_provisionedProductPlans :: Lens' ListProvisionedProductPlansResponse (Maybe [ProvisionedProductPlanSummary])

Information about the plans.

listProvisionedProductPlansResponse_httpStatus :: Lens' ListProvisionedProductPlansResponse Int

The response's http status code.