Network.AWS.ServiceCatalog.ListResourcesForTagOption

Description

Lists the resources associated with the specified TagOption.

This operation returns paginated results.

Creating a Request

data ListResourcesForTagOption

See: newListResourcesForTagOption smart constructor.

Constructors

ListResourcesForTagOption'

Fields

pageSize :: Maybe Natural

The maximum number of items to return with this call.

pageToken :: Maybe Text

The page token for the next set of results. To retrieve the first set of results, use null.

resourceType :: Maybe Text

The resource type.

Portfolio

Product

tagOptionId :: Text

The TagOption identifier.

newListResourcesForTagOption

Arguments

:: Text
$sel:tagOptionId:ListResourcesForTagOption'

-> ListResourcesForTagOption

Create a value of ListResourcesForTagOption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageSize:ListResourcesForTagOption', listResourcesForTagOption_pageSize - The maximum number of items to return with this call.

$sel:pageToken:ListResourcesForTagOption', listResourcesForTagOption_pageToken - The page token for the next set of results. To retrieve the first set of results, use null.

$sel:resourceType:ListResourcesForTagOption', listResourcesForTagOption_resourceType - The resource type.

Portfolio

Product

$sel:tagOptionId:ListResourcesForTagOption', listResourcesForTagOption_tagOptionId - The TagOption identifier.

Request Lenses

listResourcesForTagOption_pageSize :: Lens' ListResourcesForTagOption (Maybe Natural)

The maximum number of items to return with this call.

listResourcesForTagOption_pageToken :: Lens' ListResourcesForTagOption (Maybe Text)

The page token for the next set of results. To retrieve the first set of results, use null.

listResourcesForTagOption_resourceType :: Lens' ListResourcesForTagOption (Maybe Text)

The resource type.

Portfolio

Product

listResourcesForTagOption_tagOptionId :: Lens' ListResourcesForTagOption Text

The TagOption identifier.

Destructuring the Response

data ListResourcesForTagOptionResponse

See: newListResourcesForTagOptionResponse smart constructor.

Constructors

ListResourcesForTagOptionResponse'

Fields

pageToken :: Maybe Text

The page token for the next set of results. To retrieve the first set of results, use null.

resourceDetails :: Maybe [ResourceDetail]

Information about the resources.

httpStatus :: Int

The response's http status code.

newListResourcesForTagOptionResponse

Arguments

:: Int
$sel:httpStatus:ListResourcesForTagOptionResponse'

-> ListResourcesForTagOptionResponse

Create a value of ListResourcesForTagOptionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageToken:ListResourcesForTagOption', listResourcesForTagOptionResponse_pageToken - The page token for the next set of results. To retrieve the first set of results, use null.

$sel:resourceDetails:ListResourcesForTagOptionResponse', listResourcesForTagOptionResponse_resourceDetails - Information about the resources.

$sel:httpStatus:ListResourcesForTagOptionResponse', listResourcesForTagOptionResponse_httpStatus - The response's http status code.

Response Lenses

listResourcesForTagOptionResponse_pageToken :: Lens' ListResourcesForTagOptionResponse (Maybe Text)

The page token for the next set of results. To retrieve the first set of results, use null.

listResourcesForTagOptionResponse_resourceDetails :: Lens' ListResourcesForTagOptionResponse (Maybe [ResourceDetail])

Information about the resources.

listResourcesForTagOptionResponse_httpStatus :: Lens' ListResourcesForTagOptionResponse Int

The response's http status code.