Network.AWS.ServiceCatalog.ScanProvisionedProducts

Description

Lists the provisioned products that are available (not terminated).

To use additional filtering, see SearchProvisionedProducts.

This operation returns paginated results.

Creating a Request

data ScanProvisionedProducts

See: newScanProvisionedProducts smart constructor.

Constructors

ScanProvisionedProducts'

Fields

pageSize :: Maybe Natural

The maximum number of items to return with this call.

pageToken :: Maybe Text

The page token for the next set of results. To retrieve the first set of results, use null.

accessLevelFilter :: Maybe AccessLevelFilter

The access level to use to obtain results. The default is User.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

newScanProvisionedProducts :: ScanProvisionedProducts

Create a value of ScanProvisionedProducts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageSize:ScanProvisionedProducts', scanProvisionedProducts_pageSize - The maximum number of items to return with this call.

$sel:pageToken:ScanProvisionedProducts', scanProvisionedProducts_pageToken - The page token for the next set of results. To retrieve the first set of results, use null.

$sel:accessLevelFilter:ScanProvisionedProducts', scanProvisionedProducts_accessLevelFilter - The access level to use to obtain results. The default is User.

$sel:acceptLanguage:ScanProvisionedProducts', scanProvisionedProducts_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

Request Lenses

scanProvisionedProducts_pageSize :: Lens' ScanProvisionedProducts (Maybe Natural)

The maximum number of items to return with this call.

scanProvisionedProducts_pageToken :: Lens' ScanProvisionedProducts (Maybe Text)

The page token for the next set of results. To retrieve the first set of results, use null.

scanProvisionedProducts_accessLevelFilter :: Lens' ScanProvisionedProducts (Maybe AccessLevelFilter)

The access level to use to obtain results. The default is User.

scanProvisionedProducts_acceptLanguage :: Lens' ScanProvisionedProducts (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

Destructuring the Response

data ScanProvisionedProductsResponse

See: newScanProvisionedProductsResponse smart constructor.

Constructors

ScanProvisionedProductsResponse'

Fields

provisionedProducts :: Maybe [ProvisionedProductDetail]

Information about the provisioned products.

nextPageToken :: Maybe Text

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

httpStatus :: Int

The response's http status code.

newScanProvisionedProductsResponse

Arguments

:: Int
$sel:httpStatus:ScanProvisionedProductsResponse'

-> ScanProvisionedProductsResponse

Create a value of ScanProvisionedProductsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:provisionedProducts:ScanProvisionedProductsResponse', scanProvisionedProductsResponse_provisionedProducts - Information about the provisioned products.

$sel:nextPageToken:ScanProvisionedProductsResponse', scanProvisionedProductsResponse_nextPageToken - The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

$sel:httpStatus:ScanProvisionedProductsResponse', scanProvisionedProductsResponse_httpStatus - The response's http status code.

Response Lenses

scanProvisionedProductsResponse_provisionedProducts :: Lens' ScanProvisionedProductsResponse (Maybe [ProvisionedProductDetail])

Information about the provisioned products.

scanProvisionedProductsResponse_nextPageToken :: Lens' ScanProvisionedProductsResponse (Maybe Text)

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

scanProvisionedProductsResponse_httpStatus :: Lens' ScanProvisionedProductsResponse Int

The response's http status code.