Network.AWS.ServiceCatalog.SearchProductsAsAdmin

Description

Gets information about the products for the specified portfolio or all products.

This operation returns paginated results.

Creating a Request

data SearchProductsAsAdmin

See: newSearchProductsAsAdmin smart constructor.

Constructors

SearchProductsAsAdmin'

Fields

sortOrder :: Maybe SortOrder

The sort order. If no value is specified, the results are not sorted.

pageSize :: Maybe Natural

The maximum number of items to return with this call.

pageToken :: Maybe Text

The page token for the next set of results. To retrieve the first set of results, use null.

portfolioId :: Maybe Text

The portfolio identifier.

sortBy :: Maybe ProductViewSortBy

The sort field. If no value is specified, the results are not sorted.

productSource :: Maybe ProductSource

Access level of the source of the product.

filters :: Maybe (HashMap ProductViewFilterBy [Text])

The search filters. If no search filters are specified, the output includes all products to which the administrator has access.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

newSearchProductsAsAdmin :: SearchProductsAsAdmin

Create a value of SearchProductsAsAdmin with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sortOrder:SearchProductsAsAdmin', searchProductsAsAdmin_sortOrder - The sort order. If no value is specified, the results are not sorted.

$sel:pageSize:SearchProductsAsAdmin', searchProductsAsAdmin_pageSize - The maximum number of items to return with this call.

$sel:pageToken:SearchProductsAsAdmin', searchProductsAsAdmin_pageToken - The page token for the next set of results. To retrieve the first set of results, use null.

$sel:portfolioId:SearchProductsAsAdmin', searchProductsAsAdmin_portfolioId - The portfolio identifier.

$sel:sortBy:SearchProductsAsAdmin', searchProductsAsAdmin_sortBy - The sort field. If no value is specified, the results are not sorted.

$sel:productSource:SearchProductsAsAdmin', searchProductsAsAdmin_productSource - Access level of the source of the product.

$sel:filters:SearchProductsAsAdmin', searchProductsAsAdmin_filters - The search filters. If no search filters are specified, the output includes all products to which the administrator has access.

$sel:acceptLanguage:SearchProductsAsAdmin', searchProductsAsAdmin_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

Request Lenses

searchProductsAsAdmin_sortOrder :: Lens' SearchProductsAsAdmin (Maybe SortOrder)

The sort order. If no value is specified, the results are not sorted.

searchProductsAsAdmin_pageSize :: Lens' SearchProductsAsAdmin (Maybe Natural)

The maximum number of items to return with this call.

searchProductsAsAdmin_pageToken :: Lens' SearchProductsAsAdmin (Maybe Text)

The page token for the next set of results. To retrieve the first set of results, use null.

searchProductsAsAdmin_portfolioId :: Lens' SearchProductsAsAdmin (Maybe Text)

The portfolio identifier.

searchProductsAsAdmin_sortBy :: Lens' SearchProductsAsAdmin (Maybe ProductViewSortBy)

The sort field. If no value is specified, the results are not sorted.

searchProductsAsAdmin_productSource :: Lens' SearchProductsAsAdmin (Maybe ProductSource)

Access level of the source of the product.

searchProductsAsAdmin_filters :: Lens' SearchProductsAsAdmin (Maybe (HashMap ProductViewFilterBy [Text]))

The search filters. If no search filters are specified, the output includes all products to which the administrator has access.

searchProductsAsAdmin_acceptLanguage :: Lens' SearchProductsAsAdmin (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

Destructuring the Response

data SearchProductsAsAdminResponse

See: newSearchProductsAsAdminResponse smart constructor.

Constructors

SearchProductsAsAdminResponse'

Fields

nextPageToken :: Maybe Text

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

productViewDetails :: Maybe [ProductViewDetail]

Information about the product views.

httpStatus :: Int

The response's http status code.

newSearchProductsAsAdminResponse

Arguments

:: Int
$sel:httpStatus:SearchProductsAsAdminResponse'

-> SearchProductsAsAdminResponse

Create a value of SearchProductsAsAdminResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:SearchProductsAsAdminResponse', searchProductsAsAdminResponse_nextPageToken - The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

$sel:productViewDetails:SearchProductsAsAdminResponse', searchProductsAsAdminResponse_productViewDetails - Information about the product views.

$sel:httpStatus:SearchProductsAsAdminResponse', searchProductsAsAdminResponse_httpStatus - The response's http status code.

Response Lenses

searchProductsAsAdminResponse_nextPageToken :: Lens' SearchProductsAsAdminResponse (Maybe Text)

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

searchProductsAsAdminResponse_productViewDetails :: Lens' SearchProductsAsAdminResponse (Maybe [ProductViewDetail])

Information about the product views.

searchProductsAsAdminResponse_httpStatus :: Lens' SearchProductsAsAdminResponse Int

The response's http status code.