Network.AWS.ServiceCatalog.SearchProvisionedProducts

Description

Gets information about the provisioned products that meet the specified criteria.

Creating a Request

data SearchProvisionedProducts

See: newSearchProvisionedProducts smart constructor.

Constructors

SearchProvisionedProducts'

Fields

sortOrder :: Maybe SortOrder

The sort order. If no value is specified, the results are not sorted.

pageSize :: Maybe Natural

The maximum number of items to return with this call.

pageToken :: Maybe Text

The page token for the next set of results. To retrieve the first set of results, use null.

accessLevelFilter :: Maybe AccessLevelFilter

The access level to use to obtain results. The default is User.

sortBy :: Maybe Text

The sort field. If no value is specified, the results are not sorted. The valid values are arn, id, name, and lastRecordId.

filters :: Maybe (HashMap ProvisionedProductViewFilterBy [Text])

The search filters.

When the key is SearchQuery, the searchable fields are arn, createdTime, id, lastRecordId, idempotencyToken, name, physicalId, productId, provisioningArtifact, type, status, tags, userArn, userArnSession, lastProvisioningRecordId, lastSuccessfulProvisioningRecordId, productName, and provisioningArtifactName.

Example: "SearchQuery":["status:AVAILABLE"]

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

newSearchProvisionedProducts :: SearchProvisionedProducts

Create a value of SearchProvisionedProducts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sortOrder:SearchProvisionedProducts', searchProvisionedProducts_sortOrder - The sort order. If no value is specified, the results are not sorted.

$sel:pageSize:SearchProvisionedProducts', searchProvisionedProducts_pageSize - The maximum number of items to return with this call.

$sel:pageToken:SearchProvisionedProducts', searchProvisionedProducts_pageToken - The page token for the next set of results. To retrieve the first set of results, use null.

$sel:accessLevelFilter:SearchProvisionedProducts', searchProvisionedProducts_accessLevelFilter - The access level to use to obtain results. The default is User.

$sel:sortBy:SearchProvisionedProducts', searchProvisionedProducts_sortBy - The sort field. If no value is specified, the results are not sorted. The valid values are arn, id, name, and lastRecordId.

$sel:filters:SearchProvisionedProducts', searchProvisionedProducts_filters - The search filters.

When the key is SearchQuery, the searchable fields are arn, createdTime, id, lastRecordId, idempotencyToken, name, physicalId, productId, provisioningArtifact, type, status, tags, userArn, userArnSession, lastProvisioningRecordId, lastSuccessfulProvisioningRecordId, productName, and provisioningArtifactName.

Example: "SearchQuery":["status:AVAILABLE"]

$sel:acceptLanguage:SearchProvisionedProducts', searchProvisionedProducts_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

Request Lenses

searchProvisionedProducts_sortOrder :: Lens' SearchProvisionedProducts (Maybe SortOrder)

The sort order. If no value is specified, the results are not sorted.

searchProvisionedProducts_pageSize :: Lens' SearchProvisionedProducts (Maybe Natural)

The maximum number of items to return with this call.

searchProvisionedProducts_pageToken :: Lens' SearchProvisionedProducts (Maybe Text)

The page token for the next set of results. To retrieve the first set of results, use null.

searchProvisionedProducts_accessLevelFilter :: Lens' SearchProvisionedProducts (Maybe AccessLevelFilter)

The access level to use to obtain results. The default is User.

searchProvisionedProducts_sortBy :: Lens' SearchProvisionedProducts (Maybe Text)

The sort field. If no value is specified, the results are not sorted. The valid values are arn, id, name, and lastRecordId.

searchProvisionedProducts_filters :: Lens' SearchProvisionedProducts (Maybe (HashMap ProvisionedProductViewFilterBy [Text]))

The search filters.

When the key is SearchQuery, the searchable fields are arn, createdTime, id, lastRecordId, idempotencyToken, name, physicalId, productId, provisioningArtifact, type, status, tags, userArn, userArnSession, lastProvisioningRecordId, lastSuccessfulProvisioningRecordId, productName, and provisioningArtifactName.

Example: "SearchQuery":["status:AVAILABLE"]

searchProvisionedProducts_acceptLanguage :: Lens' SearchProvisionedProducts (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

Destructuring the Response

data SearchProvisionedProductsResponse

See: newSearchProvisionedProductsResponse smart constructor.

Constructors

SearchProvisionedProductsResponse'

Fields

totalResultsCount :: Maybe Int

The number of provisioned products found.

provisionedProducts :: Maybe [ProvisionedProductAttribute]

Information about the provisioned products.

nextPageToken :: Maybe Text

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

httpStatus :: Int

The response's http status code.

newSearchProvisionedProductsResponse

Arguments

:: Int
$sel:httpStatus:SearchProvisionedProductsResponse'

-> SearchProvisionedProductsResponse

Create a value of SearchProvisionedProductsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:totalResultsCount:SearchProvisionedProductsResponse', searchProvisionedProductsResponse_totalResultsCount - The number of provisioned products found.

$sel:provisionedProducts:SearchProvisionedProductsResponse', searchProvisionedProductsResponse_provisionedProducts - Information about the provisioned products.

$sel:nextPageToken:SearchProvisionedProductsResponse', searchProvisionedProductsResponse_nextPageToken - The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

$sel:httpStatus:SearchProvisionedProductsResponse', searchProvisionedProductsResponse_httpStatus - The response's http status code.

Response Lenses

searchProvisionedProductsResponse_totalResultsCount :: Lens' SearchProvisionedProductsResponse (Maybe Int)

The number of provisioned products found.

searchProvisionedProductsResponse_provisionedProducts :: Lens' SearchProvisionedProductsResponse (Maybe [ProvisionedProductAttribute])

Information about the provisioned products.

searchProvisionedProductsResponse_nextPageToken :: Lens' SearchProvisionedProductsResponse (Maybe Text)

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

searchProvisionedProductsResponse_httpStatus :: Lens' SearchProvisionedProductsResponse Int

The response's http status code.