Network.AWS.ServiceCatalog.TerminateProvisionedProduct

Description

Terminates the specified provisioned product.

This operation does not delete any records associated with the provisioned product.

You can check the status of this request using DescribeRecord.

Creating a Request

data TerminateProvisionedProduct

See: newTerminateProvisionedProduct smart constructor.

Constructors

TerminateProvisionedProduct'

Fields

provisionedProductName :: Maybe Text

The name of the provisioned product. You cannot specify both ProvisionedProductName and ProvisionedProductId.

retainPhysicalResources :: Maybe Bool

When this boolean parameter is set to true, the TerminateProvisionedProduct API deletes the Service Catalog provisioned product. However, it does not remove the CloudFormation stack, stack set, or the underlying resources of the deleted provisioned product. The default value is false.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

ignoreErrors :: Maybe Bool

If set to true, AWS Service Catalog stops managing the specified provisioned product even if it cannot delete the underlying resources.

provisionedProductId :: Maybe Text

The identifier of the provisioned product. You cannot specify both ProvisionedProductName and ProvisionedProductId.

terminateToken :: Text

An idempotency token that uniquely identifies the termination request. This token is only valid during the termination process. After the provisioned product is terminated, subsequent requests to terminate the same provisioned product always return ResourceNotFound.

newTerminateProvisionedProduct

Arguments

:: Text
$sel:terminateToken:TerminateProvisionedProduct'

-> TerminateProvisionedProduct

Create a value of TerminateProvisionedProduct with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:provisionedProductName:TerminateProvisionedProduct', terminateProvisionedProduct_provisionedProductName - The name of the provisioned product. You cannot specify both ProvisionedProductName and ProvisionedProductId.

$sel:retainPhysicalResources:TerminateProvisionedProduct', terminateProvisionedProduct_retainPhysicalResources - When this boolean parameter is set to true, the TerminateProvisionedProduct API deletes the Service Catalog provisioned product. However, it does not remove the CloudFormation stack, stack set, or the underlying resources of the deleted provisioned product. The default value is false.

$sel:acceptLanguage:TerminateProvisionedProduct', terminateProvisionedProduct_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:ignoreErrors:TerminateProvisionedProduct', terminateProvisionedProduct_ignoreErrors - If set to true, AWS Service Catalog stops managing the specified provisioned product even if it cannot delete the underlying resources.

$sel:provisionedProductId:TerminateProvisionedProduct', terminateProvisionedProduct_provisionedProductId - The identifier of the provisioned product. You cannot specify both ProvisionedProductName and ProvisionedProductId.

$sel:terminateToken:TerminateProvisionedProduct', terminateProvisionedProduct_terminateToken - An idempotency token that uniquely identifies the termination request. This token is only valid during the termination process. After the provisioned product is terminated, subsequent requests to terminate the same provisioned product always return ResourceNotFound.

Request Lenses

terminateProvisionedProduct_provisionedProductName :: Lens' TerminateProvisionedProduct (Maybe Text)

The name of the provisioned product. You cannot specify both ProvisionedProductName and ProvisionedProductId.

terminateProvisionedProduct_retainPhysicalResources :: Lens' TerminateProvisionedProduct (Maybe Bool)

When this boolean parameter is set to true, the TerminateProvisionedProduct API deletes the Service Catalog provisioned product. However, it does not remove the CloudFormation stack, stack set, or the underlying resources of the deleted provisioned product. The default value is false.

terminateProvisionedProduct_acceptLanguage :: Lens' TerminateProvisionedProduct (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

terminateProvisionedProduct_ignoreErrors :: Lens' TerminateProvisionedProduct (Maybe Bool)

If set to true, AWS Service Catalog stops managing the specified provisioned product even if it cannot delete the underlying resources.

terminateProvisionedProduct_provisionedProductId :: Lens' TerminateProvisionedProduct (Maybe Text)

The identifier of the provisioned product. You cannot specify both ProvisionedProductName and ProvisionedProductId.

terminateProvisionedProduct_terminateToken :: Lens' TerminateProvisionedProduct Text

An idempotency token that uniquely identifies the termination request. This token is only valid during the termination process. After the provisioned product is terminated, subsequent requests to terminate the same provisioned product always return ResourceNotFound.

Destructuring the Response

data TerminateProvisionedProductResponse

See: newTerminateProvisionedProductResponse smart constructor.

Constructors

TerminateProvisionedProductResponse'

Fields

recordDetail :: Maybe RecordDetail

Information about the result of this request.

httpStatus :: Int

The response's http status code.

newTerminateProvisionedProductResponse

Arguments

:: Int
$sel:httpStatus:TerminateProvisionedProductResponse'

-> TerminateProvisionedProductResponse

Create a value of TerminateProvisionedProductResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recordDetail:TerminateProvisionedProductResponse', terminateProvisionedProductResponse_recordDetail - Information about the result of this request.

$sel:httpStatus:TerminateProvisionedProductResponse', terminateProvisionedProductResponse_httpStatus - The response's http status code.

Response Lenses

terminateProvisionedProductResponse_recordDetail :: Lens' TerminateProvisionedProductResponse (Maybe RecordDetail)

Information about the result of this request.

terminateProvisionedProductResponse_httpStatus :: Lens' TerminateProvisionedProductResponse Int

The response's http status code.