Network.AWS.ServiceCatalog.Types.ConstraintDetail

Description

Documentation

data ConstraintDetail

Information about a constraint.

See: newConstraintDetail smart constructor.

Constructors

ConstraintDetail'

Fields

constraintId :: Maybe Text

The identifier of the constraint.

portfolioId :: Maybe Text

The identifier of the portfolio the product resides in. The constraint applies only to the instance of the product that lives within this portfolio.

owner :: Maybe Text

The owner of the constraint.

productId :: Maybe Text

The identifier of the product the constraint applies to. Note that a constraint applies to a specific instance of a product within a certain portfolio.

description :: Maybe Text

The description of the constraint.

type' :: Maybe Text

The type of constraint.

LAUNCH

NOTIFICATION

STACKSET

TEMPLATE

newConstraintDetail :: ConstraintDetail

Create a value of ConstraintDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:constraintId:ConstraintDetail', constraintDetail_constraintId - The identifier of the constraint.

$sel:portfolioId:ConstraintDetail', constraintDetail_portfolioId - The identifier of the portfolio the product resides in. The constraint applies only to the instance of the product that lives within this portfolio.

$sel:owner:ConstraintDetail', constraintDetail_owner - The owner of the constraint.

$sel:productId:ConstraintDetail', constraintDetail_productId - The identifier of the product the constraint applies to. Note that a constraint applies to a specific instance of a product within a certain portfolio.

$sel:description:ConstraintDetail', constraintDetail_description - The description of the constraint.

$sel:type':ConstraintDetail', constraintDetail_type - The type of constraint.

LAUNCH

NOTIFICATION

STACKSET

TEMPLATE

constraintDetail_constraintId :: Lens' ConstraintDetail (Maybe Text)

The identifier of the constraint.

constraintDetail_portfolioId :: Lens' ConstraintDetail (Maybe Text)

The identifier of the portfolio the product resides in. The constraint applies only to the instance of the product that lives within this portfolio.

constraintDetail_owner :: Lens' ConstraintDetail (Maybe Text)

The owner of the constraint.

constraintDetail_productId :: Lens' ConstraintDetail (Maybe Text)

The identifier of the product the constraint applies to. Note that a constraint applies to a specific instance of a product within a certain portfolio.

constraintDetail_description :: Lens' ConstraintDetail (Maybe Text)

The description of the constraint.

constraintDetail_type :: Lens' ConstraintDetail (Maybe Text)

The type of constraint.

LAUNCH

NOTIFICATION

STACKSET

TEMPLATE