Network.AWS.ServiceCatalog.Types.FailedServiceActionAssociation

Description

Documentation

data FailedServiceActionAssociation

An object containing information about the error, along with identifying information about the self-service action and its associations.

See: newFailedServiceActionAssociation smart constructor.

Constructors

FailedServiceActionAssociation'

Fields

provisioningArtifactId :: Maybe Text

The identifier of the provisioning artifact. For example, pa-4abcdjnxjj6ne.

productId :: Maybe Text

The product identifier. For example, prod-abcdzk7xy33qa.

serviceActionId :: Maybe Text

The self-service action identifier. For example, act-fs7abcd89wxyz.

errorMessage :: Maybe Text

A text description of the error.

errorCode :: Maybe ServiceActionAssociationErrorCode

The error code. Valid values are listed below.

newFailedServiceActionAssociation :: FailedServiceActionAssociation

Create a value of FailedServiceActionAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:provisioningArtifactId:FailedServiceActionAssociation', failedServiceActionAssociation_provisioningArtifactId - The identifier of the provisioning artifact. For example, pa-4abcdjnxjj6ne.

$sel:productId:FailedServiceActionAssociation', failedServiceActionAssociation_productId - The product identifier. For example, prod-abcdzk7xy33qa.

$sel:serviceActionId:FailedServiceActionAssociation', failedServiceActionAssociation_serviceActionId - The self-service action identifier. For example, act-fs7abcd89wxyz.

$sel:errorMessage:FailedServiceActionAssociation', failedServiceActionAssociation_errorMessage - A text description of the error.

$sel:errorCode:FailedServiceActionAssociation', failedServiceActionAssociation_errorCode - The error code. Valid values are listed below.

failedServiceActionAssociation_provisioningArtifactId :: Lens' FailedServiceActionAssociation (Maybe Text)

The identifier of the provisioning artifact. For example, pa-4abcdjnxjj6ne.

failedServiceActionAssociation_productId :: Lens' FailedServiceActionAssociation (Maybe Text)

The product identifier. For example, prod-abcdzk7xy33qa.

failedServiceActionAssociation_serviceActionId :: Lens' FailedServiceActionAssociation (Maybe Text)

The self-service action identifier. For example, act-fs7abcd89wxyz.

failedServiceActionAssociation_errorMessage :: Lens' FailedServiceActionAssociation (Maybe Text)

A text description of the error.

failedServiceActionAssociation_errorCode :: Lens' FailedServiceActionAssociation (Maybe ServiceActionAssociationErrorCode)

The error code. Valid values are listed below.