Network.AWS.ServiceCatalog.Types.LaunchPathSummary

Description

Documentation

data LaunchPathSummary

Summary information about a product path for a user.

See: newLaunchPathSummary smart constructor.

Constructors

LaunchPathSummary'

Fields

constraintSummaries :: Maybe [ConstraintSummary]

The constraints on the portfolio-product relationship.

id :: Maybe Text

The identifier of the product path.

name :: Maybe Text

The name of the portfolio to which the user was assigned.

tags :: Maybe [Tag]

The tags associated with this product path.

newLaunchPathSummary :: LaunchPathSummary

Create a value of LaunchPathSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:constraintSummaries:LaunchPathSummary', launchPathSummary_constraintSummaries - The constraints on the portfolio-product relationship.

$sel:id:LaunchPathSummary', launchPathSummary_id - The identifier of the product path.

$sel:name:LaunchPathSummary', launchPathSummary_name - The name of the portfolio to which the user was assigned.

$sel:tags:LaunchPathSummary', launchPathSummary_tags - The tags associated with this product path.

launchPathSummary_constraintSummaries :: Lens' LaunchPathSummary (Maybe [ConstraintSummary])

The constraints on the portfolio-product relationship.

launchPathSummary_id :: Lens' LaunchPathSummary (Maybe Text)

The identifier of the product path.

launchPathSummary_name :: Lens' LaunchPathSummary (Maybe Text)

The name of the portfolio to which the user was assigned.

launchPathSummary_tags :: Lens' LaunchPathSummary (Maybe [Tag])

The tags associated with this product path.