Network.AWS.ServiceCatalog.Types.PortfolioShareDetail

Description

Documentation

data PortfolioShareDetail

Information about the portfolio share.

See: newPortfolioShareDetail smart constructor.

Constructors

PortfolioShareDetail'

Fields

principalId :: Maybe Text

The identifier of the recipient entity that received the portfolio share. The recipient entities can be one of the following:

An external account.

An organziation member account.

An organzational unit (OU).

The organization itself. (This shares with every account in the organization).

shareTagOptions :: Maybe Bool

Indicates whether TagOptions sharing is enabled or disabled for the portfolio share.

type' :: Maybe DescribePortfolioShareType

The type of the portfolio share.

accepted :: Maybe Bool

Indicates whether the shared portfolio is imported by the recipient account. If the recipient is in an organization node, the share is automatically imported, and the field is always set to true.

newPortfolioShareDetail :: PortfolioShareDetail

Create a value of PortfolioShareDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:principalId:PortfolioShareDetail', portfolioShareDetail_principalId - The identifier of the recipient entity that received the portfolio share. The recipient entities can be one of the following:

An external account.

An organziation member account.

An organzational unit (OU).

The organization itself. (This shares with every account in the organization).

$sel:shareTagOptions:PortfolioShareDetail', portfolioShareDetail_shareTagOptions - Indicates whether TagOptions sharing is enabled or disabled for the portfolio share.

$sel:type':PortfolioShareDetail', portfolioShareDetail_type - The type of the portfolio share.

$sel:accepted:PortfolioShareDetail', portfolioShareDetail_accepted - Indicates whether the shared portfolio is imported by the recipient account. If the recipient is in an organization node, the share is automatically imported, and the field is always set to true.

portfolioShareDetail_principalId :: Lens' PortfolioShareDetail (Maybe Text)

The identifier of the recipient entity that received the portfolio share. The recipient entities can be one of the following:

An external account.

An organziation member account.

An organzational unit (OU).

The organization itself. (This shares with every account in the organization).

portfolioShareDetail_shareTagOptions :: Lens' PortfolioShareDetail (Maybe Bool)

Indicates whether TagOptions sharing is enabled or disabled for the portfolio share.

portfolioShareDetail_type :: Lens' PortfolioShareDetail (Maybe DescribePortfolioShareType)

The type of the portfolio share.

portfolioShareDetail_accepted :: Lens' PortfolioShareDetail (Maybe Bool)

Indicates whether the shared portfolio is imported by the recipient account. If the recipient is in an organization node, the share is automatically imported, and the field is always set to true.