Network.AWS.ServiceCatalog.Types.ProvisionedProductAttribute

Description

Documentation

data ProvisionedProductAttribute

Information about a provisioned product.

See: newProvisionedProductAttribute smart constructor.

Constructors

ProvisionedProductAttribute'

Fields

lastSuccessfulProvisioningRecordId :: Maybe Text

The record identifier of the last successful request performed on this provisioned product of the following types:

ProvisionedProduct

UpdateProvisionedProduct

ExecuteProvisionedProductPlan

TerminateProvisionedProduct

statusMessage :: Maybe Text

The current status message of the provisioned product.

status :: Maybe ProvisionedProductStatus

The current status of the provisioned product.

AVAILABLE - Stable state, ready to perform any operation. The most recent operation succeeded and completed.

UNDER_CHANGE - Transitive state. Operations performed might not have valid results. Wait for an AVAILABLE status before performing operations.

TAINTED - Stable state, ready to perform any operation. The stack has completed the requested operation but is not exactly what was requested. For example, a request to update to a new version failed and the stack rolled back to the current version.

ERROR - An unexpected error occurred. The provisioned product exists but the stack is not running. For example, CloudFormation received a parameter value that was not valid and could not launch the stack.

PLAN_IN_PROGRESS - Transitive state. The plan operations were performed to provision a new product, but resources have not yet been created. After reviewing the list of resources to be created, execute the plan. Wait for an AVAILABLE status before performing operations.

idempotencyToken :: Maybe Text

A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

userArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM user.

provisioningArtifactName :: Maybe Text

The name of the provisioning artifact.

arn :: Maybe Text

The ARN of the provisioned product.

id :: Maybe Text

The identifier of the provisioned product.

createdTime :: Maybe POSIX

The UTC time stamp of the creation time.

provisioningArtifactId :: Maybe Text

The identifier of the provisioning artifact.

name :: Maybe Text

The user-friendly name of the provisioned product.

productName :: Maybe Text

The name of the product.

tags :: Maybe [Tag]

One or more tags.

productId :: Maybe Text

The product identifier.

lastProvisioningRecordId :: Maybe Text

The record identifier of the last request performed on this provisioned product of the following types:

ProvisionedProduct

UpdateProvisionedProduct

ExecuteProvisionedProductPlan

TerminateProvisionedProduct

physicalId :: Maybe Text

The assigned identifier for the resource, such as an EC2 instance ID or an S3 bucket name.

type' :: Maybe Text

The type of provisioned product. The supported values are CFN_STACK and CFN_STACKSET.

lastRecordId :: Maybe Text

The record identifier of the last request performed on this provisioned product.

userArnSession :: Maybe Text

The ARN of the IAM user in the session. This ARN might contain a session ID.

newProvisionedProductAttribute :: ProvisionedProductAttribute

Create a value of ProvisionedProductAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastSuccessfulProvisioningRecordId:ProvisionedProductAttribute', provisionedProductAttribute_lastSuccessfulProvisioningRecordId - The record identifier of the last successful request performed on this provisioned product of the following types:

ProvisionedProduct

UpdateProvisionedProduct

ExecuteProvisionedProductPlan

TerminateProvisionedProduct

$sel:statusMessage:ProvisionedProductAttribute', provisionedProductAttribute_statusMessage - The current status message of the provisioned product.

$sel:status:ProvisionedProductAttribute', provisionedProductAttribute_status - The current status of the provisioned product.

AVAILABLE - Stable state, ready to perform any operation. The most recent operation succeeded and completed.

UNDER_CHANGE - Transitive state. Operations performed might not have valid results. Wait for an AVAILABLE status before performing operations.

TAINTED - Stable state, ready to perform any operation. The stack has completed the requested operation but is not exactly what was requested. For example, a request to update to a new version failed and the stack rolled back to the current version.

ERROR - An unexpected error occurred. The provisioned product exists but the stack is not running. For example, CloudFormation received a parameter value that was not valid and could not launch the stack.

PLAN_IN_PROGRESS - Transitive state. The plan operations were performed to provision a new product, but resources have not yet been created. After reviewing the list of resources to be created, execute the plan. Wait for an AVAILABLE status before performing operations.

$sel:idempotencyToken:ProvisionedProductAttribute', provisionedProductAttribute_idempotencyToken - A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

$sel:userArn:ProvisionedProductAttribute', provisionedProductAttribute_userArn - The Amazon Resource Name (ARN) of the IAM user.

$sel:provisioningArtifactName:ProvisionedProductAttribute', provisionedProductAttribute_provisioningArtifactName - The name of the provisioning artifact.

$sel:arn:ProvisionedProductAttribute', provisionedProductAttribute_arn - The ARN of the provisioned product.

$sel:id:ProvisionedProductAttribute', provisionedProductAttribute_id - The identifier of the provisioned product.

$sel:createdTime:ProvisionedProductAttribute', provisionedProductAttribute_createdTime - The UTC time stamp of the creation time.

$sel:provisioningArtifactId:ProvisionedProductAttribute', provisionedProductAttribute_provisioningArtifactId - The identifier of the provisioning artifact.

$sel:name:ProvisionedProductAttribute', provisionedProductAttribute_name - The user-friendly name of the provisioned product.

$sel:productName:ProvisionedProductAttribute', provisionedProductAttribute_productName - The name of the product.

$sel:tags:ProvisionedProductAttribute', provisionedProductAttribute_tags - One or more tags.

$sel:productId:ProvisionedProductAttribute', provisionedProductAttribute_productId - The product identifier.

$sel:lastProvisioningRecordId:ProvisionedProductAttribute', provisionedProductAttribute_lastProvisioningRecordId - The record identifier of the last request performed on this provisioned product of the following types:

ProvisionedProduct

UpdateProvisionedProduct

ExecuteProvisionedProductPlan

TerminateProvisionedProduct

$sel:physicalId:ProvisionedProductAttribute', provisionedProductAttribute_physicalId - The assigned identifier for the resource, such as an EC2 instance ID or an S3 bucket name.

$sel:type':ProvisionedProductAttribute', provisionedProductAttribute_type - The type of provisioned product. The supported values are CFN_STACK and CFN_STACKSET.

$sel:lastRecordId:ProvisionedProductAttribute', provisionedProductAttribute_lastRecordId - The record identifier of the last request performed on this provisioned product.

$sel:userArnSession:ProvisionedProductAttribute', provisionedProductAttribute_userArnSession - The ARN of the IAM user in the session. This ARN might contain a session ID.

provisionedProductAttribute_lastSuccessfulProvisioningRecordId :: Lens' ProvisionedProductAttribute (Maybe Text)

The record identifier of the last successful request performed on this provisioned product of the following types:

ProvisionedProduct

UpdateProvisionedProduct

ExecuteProvisionedProductPlan

TerminateProvisionedProduct

provisionedProductAttribute_statusMessage :: Lens' ProvisionedProductAttribute (Maybe Text)

The current status message of the provisioned product.

provisionedProductAttribute_status :: Lens' ProvisionedProductAttribute (Maybe ProvisionedProductStatus)

The current status of the provisioned product.

AVAILABLE - Stable state, ready to perform any operation. The most recent operation succeeded and completed.

UNDER_CHANGE - Transitive state. Operations performed might not have valid results. Wait for an AVAILABLE status before performing operations.

TAINTED - Stable state, ready to perform any operation. The stack has completed the requested operation but is not exactly what was requested. For example, a request to update to a new version failed and the stack rolled back to the current version.

ERROR - An unexpected error occurred. The provisioned product exists but the stack is not running. For example, CloudFormation received a parameter value that was not valid and could not launch the stack.

PLAN_IN_PROGRESS - Transitive state. The plan operations were performed to provision a new product, but resources have not yet been created. After reviewing the list of resources to be created, execute the plan. Wait for an AVAILABLE status before performing operations.

provisionedProductAttribute_idempotencyToken :: Lens' ProvisionedProductAttribute (Maybe Text)

A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

provisionedProductAttribute_userArn :: Lens' ProvisionedProductAttribute (Maybe Text)

The Amazon Resource Name (ARN) of the IAM user.

provisionedProductAttribute_provisioningArtifactName :: Lens' ProvisionedProductAttribute (Maybe Text)

The name of the provisioning artifact.

provisionedProductAttribute_arn :: Lens' ProvisionedProductAttribute (Maybe Text)

The ARN of the provisioned product.

provisionedProductAttribute_id :: Lens' ProvisionedProductAttribute (Maybe Text)

The identifier of the provisioned product.

provisionedProductAttribute_createdTime :: Lens' ProvisionedProductAttribute (Maybe UTCTime)

The UTC time stamp of the creation time.

provisionedProductAttribute_provisioningArtifactId :: Lens' ProvisionedProductAttribute (Maybe Text)

The identifier of the provisioning artifact.

provisionedProductAttribute_name :: Lens' ProvisionedProductAttribute (Maybe Text)

The user-friendly name of the provisioned product.

provisionedProductAttribute_productName :: Lens' ProvisionedProductAttribute (Maybe Text)

The name of the product.

provisionedProductAttribute_tags :: Lens' ProvisionedProductAttribute (Maybe [Tag])

One or more tags.

provisionedProductAttribute_productId :: Lens' ProvisionedProductAttribute (Maybe Text)

The product identifier.

provisionedProductAttribute_lastProvisioningRecordId :: Lens' ProvisionedProductAttribute (Maybe Text)

The record identifier of the last request performed on this provisioned product of the following types:

ProvisionedProduct

UpdateProvisionedProduct

ExecuteProvisionedProductPlan

TerminateProvisionedProduct

provisionedProductAttribute_physicalId :: Lens' ProvisionedProductAttribute (Maybe Text)

The assigned identifier for the resource, such as an EC2 instance ID or an S3 bucket name.

provisionedProductAttribute_type :: Lens' ProvisionedProductAttribute (Maybe Text)

The type of provisioned product. The supported values are CFN_STACK and CFN_STACKSET.

provisionedProductAttribute_lastRecordId :: Lens' ProvisionedProductAttribute (Maybe Text)

The record identifier of the last request performed on this provisioned product.

provisionedProductAttribute_userArnSession :: Lens' ProvisionedProductAttribute (Maybe Text)

The ARN of the IAM user in the session. This ARN might contain a session ID.