Network.AWS.ServiceCatalog.Types.ProvisionedProductDetail

Description

Documentation

data ProvisionedProductDetail

Information about a provisioned product.

See: newProvisionedProductDetail smart constructor.

Constructors

ProvisionedProductDetail'

Fields

lastSuccessfulProvisioningRecordId :: Maybe Text

The record identifier of the last successful request performed on this provisioned product of the following types:

ProvisionedProduct

UpdateProvisionedProduct

ExecuteProvisionedProductPlan

TerminateProvisionedProduct

statusMessage :: Maybe Text

The current status message of the provisioned product.

status :: Maybe ProvisionedProductStatus

The current status of the provisioned product.

AVAILABLE - Stable state, ready to perform any operation. The most recent operation succeeded and completed.

UNDER_CHANGE - Transitive state. Operations performed might not have valid results. Wait for an AVAILABLE status before performing operations.

TAINTED - Stable state, ready to perform any operation. The stack has completed the requested operation but is not exactly what was requested. For example, a request to update to a new version failed and the stack rolled back to the current version.

ERROR - An unexpected error occurred. The provisioned product exists but the stack is not running. For example, CloudFormation received a parameter value that was not valid and could not launch the stack.

PLAN_IN_PROGRESS - Transitive state. The plan operations were performed to provision a new product, but resources have not yet been created. After reviewing the list of resources to be created, execute the plan. Wait for an AVAILABLE status before performing operations.

idempotencyToken :: Maybe Text

A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

arn :: Maybe Text

The ARN of the provisioned product.

id :: Maybe Text

The identifier of the provisioned product.

createdTime :: Maybe POSIX

The UTC time stamp of the creation time.

provisioningArtifactId :: Maybe Text

The identifier of the provisioning artifact. For example, pa-4abcdjnxjj6ne.

name :: Maybe Text

The user-friendly name of the provisioned product.

launchRoleArn :: Maybe Text

The ARN of the launch role associated with the provisioned product.

productId :: Maybe Text

The product identifier. For example, prod-abcdzk7xy33qa.

lastProvisioningRecordId :: Maybe Text

The record identifier of the last request performed on this provisioned product of the following types:

ProvisionedProduct

UpdateProvisionedProduct

ExecuteProvisionedProductPlan

TerminateProvisionedProduct

type' :: Maybe Text

The type of provisioned product. The supported values are CFN_STACK and CFN_STACKSET.

lastRecordId :: Maybe Text

The record identifier of the last request performed on this provisioned product.

newProvisionedProductDetail :: ProvisionedProductDetail

Create a value of ProvisionedProductDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastSuccessfulProvisioningRecordId:ProvisionedProductDetail', provisionedProductDetail_lastSuccessfulProvisioningRecordId - The record identifier of the last successful request performed on this provisioned product of the following types:

ProvisionedProduct

UpdateProvisionedProduct

ExecuteProvisionedProductPlan

TerminateProvisionedProduct

$sel:statusMessage:ProvisionedProductDetail', provisionedProductDetail_statusMessage - The current status message of the provisioned product.

$sel:status:ProvisionedProductDetail', provisionedProductDetail_status - The current status of the provisioned product.

AVAILABLE - Stable state, ready to perform any operation. The most recent operation succeeded and completed.

UNDER_CHANGE - Transitive state. Operations performed might not have valid results. Wait for an AVAILABLE status before performing operations.

TAINTED - Stable state, ready to perform any operation. The stack has completed the requested operation but is not exactly what was requested. For example, a request to update to a new version failed and the stack rolled back to the current version.

ERROR - An unexpected error occurred. The provisioned product exists but the stack is not running. For example, CloudFormation received a parameter value that was not valid and could not launch the stack.

PLAN_IN_PROGRESS - Transitive state. The plan operations were performed to provision a new product, but resources have not yet been created. After reviewing the list of resources to be created, execute the plan. Wait for an AVAILABLE status before performing operations.

$sel:idempotencyToken:ProvisionedProductDetail', provisionedProductDetail_idempotencyToken - A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

$sel:arn:ProvisionedProductDetail', provisionedProductDetail_arn - The ARN of the provisioned product.

$sel:id:ProvisionedProductDetail', provisionedProductDetail_id - The identifier of the provisioned product.

$sel:createdTime:ProvisionedProductDetail', provisionedProductDetail_createdTime - The UTC time stamp of the creation time.

$sel:provisioningArtifactId:ProvisionedProductDetail', provisionedProductDetail_provisioningArtifactId - The identifier of the provisioning artifact. For example, pa-4abcdjnxjj6ne.

$sel:name:ProvisionedProductDetail', provisionedProductDetail_name - The user-friendly name of the provisioned product.

$sel:launchRoleArn:ProvisionedProductDetail', provisionedProductDetail_launchRoleArn - The ARN of the launch role associated with the provisioned product.

$sel:productId:ProvisionedProductDetail', provisionedProductDetail_productId - The product identifier. For example, prod-abcdzk7xy33qa.

$sel:lastProvisioningRecordId:ProvisionedProductDetail', provisionedProductDetail_lastProvisioningRecordId - The record identifier of the last request performed on this provisioned product of the following types:

ProvisionedProduct

UpdateProvisionedProduct

ExecuteProvisionedProductPlan

TerminateProvisionedProduct

$sel:type':ProvisionedProductDetail', provisionedProductDetail_type - The type of provisioned product. The supported values are CFN_STACK and CFN_STACKSET.

$sel:lastRecordId:ProvisionedProductDetail', provisionedProductDetail_lastRecordId - The record identifier of the last request performed on this provisioned product.

provisionedProductDetail_lastSuccessfulProvisioningRecordId :: Lens' ProvisionedProductDetail (Maybe Text)

The record identifier of the last successful request performed on this provisioned product of the following types:

ProvisionedProduct

UpdateProvisionedProduct

ExecuteProvisionedProductPlan

TerminateProvisionedProduct

provisionedProductDetail_statusMessage :: Lens' ProvisionedProductDetail (Maybe Text)

The current status message of the provisioned product.

provisionedProductDetail_status :: Lens' ProvisionedProductDetail (Maybe ProvisionedProductStatus)

The current status of the provisioned product.

AVAILABLE - Stable state, ready to perform any operation. The most recent operation succeeded and completed.

UNDER_CHANGE - Transitive state. Operations performed might not have valid results. Wait for an AVAILABLE status before performing operations.

TAINTED - Stable state, ready to perform any operation. The stack has completed the requested operation but is not exactly what was requested. For example, a request to update to a new version failed and the stack rolled back to the current version.

ERROR - An unexpected error occurred. The provisioned product exists but the stack is not running. For example, CloudFormation received a parameter value that was not valid and could not launch the stack.

PLAN_IN_PROGRESS - Transitive state. The plan operations were performed to provision a new product, but resources have not yet been created. After reviewing the list of resources to be created, execute the plan. Wait for an AVAILABLE status before performing operations.

provisionedProductDetail_idempotencyToken :: Lens' ProvisionedProductDetail (Maybe Text)

A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

provisionedProductDetail_arn :: Lens' ProvisionedProductDetail (Maybe Text)

The ARN of the provisioned product.

provisionedProductDetail_id :: Lens' ProvisionedProductDetail (Maybe Text)

The identifier of the provisioned product.

provisionedProductDetail_createdTime :: Lens' ProvisionedProductDetail (Maybe UTCTime)

The UTC time stamp of the creation time.

provisionedProductDetail_provisioningArtifactId :: Lens' ProvisionedProductDetail (Maybe Text)

The identifier of the provisioning artifact. For example, pa-4abcdjnxjj6ne.

provisionedProductDetail_name :: Lens' ProvisionedProductDetail (Maybe Text)

The user-friendly name of the provisioned product.

provisionedProductDetail_launchRoleArn :: Lens' ProvisionedProductDetail (Maybe Text)

The ARN of the launch role associated with the provisioned product.

provisionedProductDetail_productId :: Lens' ProvisionedProductDetail (Maybe Text)

The product identifier. For example, prod-abcdzk7xy33qa.

provisionedProductDetail_lastProvisioningRecordId :: Lens' ProvisionedProductDetail (Maybe Text)

The record identifier of the last request performed on this provisioned product of the following types:

ProvisionedProduct

UpdateProvisionedProduct

ExecuteProvisionedProductPlan

TerminateProvisionedProduct

provisionedProductDetail_type :: Lens' ProvisionedProductDetail (Maybe Text)

The type of provisioned product. The supported values are CFN_STACK and CFN_STACKSET.

provisionedProductDetail_lastRecordId :: Lens' ProvisionedProductDetail (Maybe Text)

The record identifier of the last request performed on this provisioned product.