Network.AWS.ServiceCatalog.Types.ProvisioningArtifactPreferences

Description

Documentation

data ProvisioningArtifactPreferences

The user-defined preferences that will be applied during product provisioning, unless overridden by ProvisioningPreferences or UpdateProvisioningPreferences.

For more information on maximum concurrent accounts and failure tolerance, see Stack set operation options in the AWS CloudFormation User Guide.

See: newProvisioningArtifactPreferences smart constructor.

Constructors

ProvisioningArtifactPreferences'

Fields

stackSetAccounts :: Maybe [Text]

One or more AWS accounts where stack instances are deployed from the stack set. These accounts can be scoped in ProvisioningPreferences$StackSetAccounts and UpdateProvisioningPreferences$StackSetAccounts.

Applicable only to a CFN_STACKSET provisioned product type.

stackSetRegions :: Maybe [Text]

One or more AWS Regions where stack instances are deployed from the stack set. These regions can be scoped in ProvisioningPreferences$StackSetRegions and UpdateProvisioningPreferences$StackSetRegions.

Applicable only to a CFN_STACKSET provisioned product type.

newProvisioningArtifactPreferences :: ProvisioningArtifactPreferences

Create a value of ProvisioningArtifactPreferences with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackSetAccounts:ProvisioningArtifactPreferences', provisioningArtifactPreferences_stackSetAccounts - One or more AWS accounts where stack instances are deployed from the stack set. These accounts can be scoped in ProvisioningPreferences$StackSetAccounts and UpdateProvisioningPreferences$StackSetAccounts.

Applicable only to a CFN_STACKSET provisioned product type.

$sel:stackSetRegions:ProvisioningArtifactPreferences', provisioningArtifactPreferences_stackSetRegions - One or more AWS Regions where stack instances are deployed from the stack set. These regions can be scoped in ProvisioningPreferences$StackSetRegions and UpdateProvisioningPreferences$StackSetRegions.

Applicable only to a CFN_STACKSET provisioned product type.

provisioningArtifactPreferences_stackSetAccounts :: Lens' ProvisioningArtifactPreferences (Maybe [Text])

One or more AWS accounts where stack instances are deployed from the stack set. These accounts can be scoped in ProvisioningPreferences$StackSetAccounts and UpdateProvisioningPreferences$StackSetAccounts.

Applicable only to a CFN_STACKSET provisioned product type.

provisioningArtifactPreferences_stackSetRegions :: Lens' ProvisioningArtifactPreferences (Maybe [Text])

One or more AWS Regions where stack instances are deployed from the stack set. These regions can be scoped in ProvisioningPreferences$StackSetRegions and UpdateProvisioningPreferences$StackSetRegions.

Applicable only to a CFN_STACKSET provisioned product type.