Network.AWS.ServiceCatalog.Types.ProvisioningArtifactProperties

Description

Documentation

data ProvisioningArtifactProperties

Information about a provisioning artifact (also known as a version) for a product.

See: newProvisioningArtifactProperties smart constructor.

Constructors

ProvisioningArtifactProperties'

Fields

disableTemplateValidation :: Maybe Bool

If set to true, AWS Service Catalog stops validating the specified provisioning artifact even if it is invalid.

name :: Maybe Text

The name of the provisioning artifact (for example, v1 v2beta). No spaces are allowed.

description :: Maybe Text

The description of the provisioning artifact, including how it differs from the previous provisioning artifact.

type' :: Maybe ProvisioningArtifactType

The type of provisioning artifact.

CLOUD_FORMATION_TEMPLATE - AWS CloudFormation template

MARKETPLACE_AMI - AWS Marketplace AMI

MARKETPLACE_CAR - AWS Marketplace Clusters and AWS Resources

info :: HashMap Text Text

Specify the template source with one of the following options, but not both. Keys accepted: [ LoadTemplateFromURL, ImportFromPhysicalId ]

The URL of the CloudFormation template in Amazon S3. Specify the URL in JSON format as follows:

"LoadTemplateFromURL": "https://s3.amazonaws.com/cf-templates-ozkq9d3hgiq2-us-east-1/..."

ImportFromPhysicalId: The physical id of the resource that contains the template. Currently only supports CloudFormation stack arn. Specify the physical id in JSON format as follows: ImportFromPhysicalId: “arn:aws:cloudformation:[us-east-1]:[accountId]:stack/[StackName]/[resourceId]

newProvisioningArtifactProperties :: ProvisioningArtifactProperties

Create a value of ProvisioningArtifactProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:disableTemplateValidation:ProvisioningArtifactProperties', provisioningArtifactProperties_disableTemplateValidation - If set to true, AWS Service Catalog stops validating the specified provisioning artifact even if it is invalid.

$sel:name:ProvisioningArtifactProperties', provisioningArtifactProperties_name - The name of the provisioning artifact (for example, v1 v2beta). No spaces are allowed.

$sel:description:ProvisioningArtifactProperties', provisioningArtifactProperties_description - The description of the provisioning artifact, including how it differs from the previous provisioning artifact.

$sel:type':ProvisioningArtifactProperties', provisioningArtifactProperties_type - The type of provisioning artifact.

CLOUD_FORMATION_TEMPLATE - AWS CloudFormation template

MARKETPLACE_AMI - AWS Marketplace AMI

MARKETPLACE_CAR - AWS Marketplace Clusters and AWS Resources

$sel:info:ProvisioningArtifactProperties', provisioningArtifactProperties_info - Specify the template source with one of the following options, but not both. Keys accepted: [ LoadTemplateFromURL, ImportFromPhysicalId ]

The URL of the CloudFormation template in Amazon S3. Specify the URL in JSON format as follows:

"LoadTemplateFromURL": "https://s3.amazonaws.com/cf-templates-ozkq9d3hgiq2-us-east-1/..."

ImportFromPhysicalId: The physical id of the resource that contains the template. Currently only supports CloudFormation stack arn. Specify the physical id in JSON format as follows: ImportFromPhysicalId: “arn:aws:cloudformation:[us-east-1]:[accountId]:stack/[StackName]/[resourceId]

provisioningArtifactProperties_disableTemplateValidation :: Lens' ProvisioningArtifactProperties (Maybe Bool)

If set to true, AWS Service Catalog stops validating the specified provisioning artifact even if it is invalid.

provisioningArtifactProperties_name :: Lens' ProvisioningArtifactProperties (Maybe Text)

The name of the provisioning artifact (for example, v1 v2beta). No spaces are allowed.

provisioningArtifactProperties_description :: Lens' ProvisioningArtifactProperties (Maybe Text)

The description of the provisioning artifact, including how it differs from the previous provisioning artifact.

provisioningArtifactProperties_type :: Lens' ProvisioningArtifactProperties (Maybe ProvisioningArtifactType)

The type of provisioning artifact.

CLOUD_FORMATION_TEMPLATE - AWS CloudFormation template

MARKETPLACE_AMI - AWS Marketplace AMI

MARKETPLACE_CAR - AWS Marketplace Clusters and AWS Resources

provisioningArtifactProperties_info :: Lens' ProvisioningArtifactProperties (HashMap Text Text)

Specify the template source with one of the following options, but not both. Keys accepted: [ LoadTemplateFromURL, ImportFromPhysicalId ]

The URL of the CloudFormation template in Amazon S3. Specify the URL in JSON format as follows:

"LoadTemplateFromURL": "https://s3.amazonaws.com/cf-templates-ozkq9d3hgiq2-us-east-1/..."

ImportFromPhysicalId: The physical id of the resource that contains the template. Currently only supports CloudFormation stack arn. Specify the physical id in JSON format as follows: ImportFromPhysicalId: “arn:aws:cloudformation:[us-east-1]:[accountId]:stack/[StackName]/[resourceId]