Network.AWS.ServiceCatalog.Types.ProvisioningArtifactView

Description

Documentation

data ProvisioningArtifactView

An object that contains summary information about a product view and a provisioning artifact.

See: newProvisioningArtifactView smart constructor.

Constructors

ProvisioningArtifactView'

Fields

productViewSummary :: Maybe ProductViewSummary

Summary information about a product view.

provisioningArtifact :: Maybe ProvisioningArtifact

Information about a provisioning artifact. A provisioning artifact is also known as a product version.

newProvisioningArtifactView :: ProvisioningArtifactView

Create a value of ProvisioningArtifactView with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:productViewSummary:ProvisioningArtifactView', provisioningArtifactView_productViewSummary - Summary information about a product view.

$sel:provisioningArtifact:ProvisioningArtifactView', provisioningArtifactView_provisioningArtifact - Information about a provisioning artifact. A provisioning artifact is also known as a product version.

provisioningArtifactView_productViewSummary :: Lens' ProvisioningArtifactView (Maybe ProductViewSummary)

Summary information about a product view.

provisioningArtifactView_provisioningArtifact :: Lens' ProvisioningArtifactView (Maybe ProvisioningArtifact)

Information about a provisioning artifact. A provisioning artifact is also known as a product version.