Network.AWS.ServiceCatalog.Types.ProvisioningParameter

Description

Documentation

data ProvisioningParameter

Information about a parameter used to provision a product.

See: newProvisioningParameter smart constructor.

Constructors

ProvisioningParameter'

Fields

key :: Maybe Text

The parameter key.

value :: Maybe Text

The parameter value.

newProvisioningParameter :: ProvisioningParameter

Create a value of ProvisioningParameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:ProvisioningParameter', provisioningParameter_key - The parameter key.

$sel:value:ProvisioningParameter', provisioningParameter_value - The parameter value.

provisioningParameter_key :: Lens' ProvisioningParameter (Maybe Text)

The parameter key.

provisioningParameter_value :: Lens' ProvisioningParameter (Maybe Text)

The parameter value.