Network.AWS.ServiceCatalog.Types.RecordDetail

Description

Documentation

data RecordDetail

Information about a request operation.

See: newRecordDetail smart constructor.

Constructors

RecordDetail'

Fields

launchRoleArn :: Maybe Text

The ARN of the launch role associated with the provisioned product.

status :: Maybe RecordStatus

The status of the provisioned product.

CREATED - The request was created but the operation has not started.

IN_PROGRESS - The requested operation is in progress.

IN_PROGRESS_IN_ERROR - The provisioned product is under change but the requested operation failed and some remediation is occurring. For example, a rollback.

SUCCEEDED - The requested operation has successfully completed.

FAILED - The requested operation has unsuccessfully completed. Investigate using the error messages returned.

recordTags :: Maybe [RecordTag]

One or more tags.

provisionedProductName :: Maybe Text

The user-friendly name of the provisioned product.

provisioningArtifactId :: Maybe Text

The identifier of the provisioning artifact.

createdTime :: Maybe POSIX

The UTC time stamp of the creation time.

recordType :: Maybe Text

The record type.

PROVISION_PRODUCT

UPDATE_PROVISIONED_PRODUCT

TERMINATE_PROVISIONED_PRODUCT

recordId :: Maybe Text

The identifier of the record.

provisionedProductType :: Maybe Text

The type of provisioned product. The supported values are CFN_STACK and CFN_STACKSET.

updatedTime :: Maybe POSIX

The time when the record was last updated.

pathId :: Maybe Text

The path identifier.

provisionedProductId :: Maybe Text

The identifier of the provisioned product.

recordErrors :: Maybe [RecordError]

The errors that occurred.

productId :: Maybe Text

The product identifier.

newRecordDetail :: RecordDetail

Create a value of RecordDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchRoleArn:RecordDetail', recordDetail_launchRoleArn - The ARN of the launch role associated with the provisioned product.

$sel:status:RecordDetail', recordDetail_status - The status of the provisioned product.

CREATED - The request was created but the operation has not started.

IN_PROGRESS - The requested operation is in progress.

IN_PROGRESS_IN_ERROR - The provisioned product is under change but the requested operation failed and some remediation is occurring. For example, a rollback.

SUCCEEDED - The requested operation has successfully completed.

FAILED - The requested operation has unsuccessfully completed. Investigate using the error messages returned.

$sel:recordTags:RecordDetail', recordDetail_recordTags - One or more tags.

$sel:provisionedProductName:RecordDetail', recordDetail_provisionedProductName - The user-friendly name of the provisioned product.

$sel:provisioningArtifactId:RecordDetail', recordDetail_provisioningArtifactId - The identifier of the provisioning artifact.

$sel:createdTime:RecordDetail', recordDetail_createdTime - The UTC time stamp of the creation time.

$sel:recordType:RecordDetail', recordDetail_recordType - The record type.

PROVISION_PRODUCT

UPDATE_PROVISIONED_PRODUCT

TERMINATE_PROVISIONED_PRODUCT

$sel:recordId:RecordDetail', recordDetail_recordId - The identifier of the record.

$sel:provisionedProductType:RecordDetail', recordDetail_provisionedProductType - The type of provisioned product. The supported values are CFN_STACK and CFN_STACKSET.

$sel:updatedTime:RecordDetail', recordDetail_updatedTime - The time when the record was last updated.

$sel:pathId:RecordDetail', recordDetail_pathId - The path identifier.

$sel:provisionedProductId:RecordDetail', recordDetail_provisionedProductId - The identifier of the provisioned product.

$sel:recordErrors:RecordDetail', recordDetail_recordErrors - The errors that occurred.

$sel:productId:RecordDetail', recordDetail_productId - The product identifier.

recordDetail_launchRoleArn :: Lens' RecordDetail (Maybe Text)

The ARN of the launch role associated with the provisioned product.

recordDetail_status :: Lens' RecordDetail (Maybe RecordStatus)

The status of the provisioned product.

CREATED - The request was created but the operation has not started.

IN_PROGRESS - The requested operation is in progress.

IN_PROGRESS_IN_ERROR - The provisioned product is under change but the requested operation failed and some remediation is occurring. For example, a rollback.

SUCCEEDED - The requested operation has successfully completed.

FAILED - The requested operation has unsuccessfully completed. Investigate using the error messages returned.

recordDetail_recordTags :: Lens' RecordDetail (Maybe [RecordTag])

One or more tags.

recordDetail_provisionedProductName :: Lens' RecordDetail (Maybe Text)

The user-friendly name of the provisioned product.

recordDetail_provisioningArtifactId :: Lens' RecordDetail (Maybe Text)

The identifier of the provisioning artifact.

recordDetail_createdTime :: Lens' RecordDetail (Maybe UTCTime)

The UTC time stamp of the creation time.

recordDetail_recordType :: Lens' RecordDetail (Maybe Text)

The record type.

PROVISION_PRODUCT

UPDATE_PROVISIONED_PRODUCT

TERMINATE_PROVISIONED_PRODUCT

recordDetail_recordId :: Lens' RecordDetail (Maybe Text)

The identifier of the record.

recordDetail_provisionedProductType :: Lens' RecordDetail (Maybe Text)

The type of provisioned product. The supported values are CFN_STACK and CFN_STACKSET.

recordDetail_updatedTime :: Lens' RecordDetail (Maybe UTCTime)

The time when the record was last updated.

recordDetail_pathId :: Lens' RecordDetail (Maybe Text)

The path identifier.

recordDetail_provisionedProductId :: Lens' RecordDetail (Maybe Text)

The identifier of the provisioned product.

recordDetail_recordErrors :: Lens' RecordDetail (Maybe [RecordError])

The errors that occurred.

recordDetail_productId :: Lens' RecordDetail (Maybe Text)

The product identifier.