Network.AWS.ServiceCatalog.Types.RecordError

Description

Documentation

data RecordError

The error code and description resulting from an operation.

See: newRecordError smart constructor.

Constructors

RecordError'

Fields

code :: Maybe Text

The numeric value of the error.

description :: Maybe Text

The description of the error.

newRecordError :: RecordError

Create a value of RecordError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:code:RecordError', recordError_code - The numeric value of the error.

$sel:description:RecordError', recordError_description - The description of the error.

recordError_code :: Lens' RecordError (Maybe Text)

The numeric value of the error.

recordError_description :: Lens' RecordError (Maybe Text)

The description of the error.