Network.AWS.ServiceCatalog.Types.ResourceChangeDetail

Description

Documentation

data ResourceChangeDetail

Information about a change to a resource attribute.

See: newResourceChangeDetail smart constructor.

Constructors

ResourceChangeDetail'

Fields

causingEntity :: Maybe Text

The ID of the entity that caused the change.

evaluation :: Maybe EvaluationType

For static evaluations, the value of the resource attribute will change and the new value is known. For dynamic evaluations, the value might change, and any new value will be determined when the plan is updated.

target :: Maybe ResourceTargetDefinition

Information about the resource attribute to be modified.

newResourceChangeDetail :: ResourceChangeDetail

Create a value of ResourceChangeDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:causingEntity:ResourceChangeDetail', resourceChangeDetail_causingEntity - The ID of the entity that caused the change.

$sel:evaluation:ResourceChangeDetail', resourceChangeDetail_evaluation - For static evaluations, the value of the resource attribute will change and the new value is known. For dynamic evaluations, the value might change, and any new value will be determined when the plan is updated.

$sel:target:ResourceChangeDetail', resourceChangeDetail_target - Information about the resource attribute to be modified.

resourceChangeDetail_causingEntity :: Lens' ResourceChangeDetail (Maybe Text)

The ID of the entity that caused the change.

resourceChangeDetail_evaluation :: Lens' ResourceChangeDetail (Maybe EvaluationType)

For static evaluations, the value of the resource attribute will change and the new value is known. For dynamic evaluations, the value might change, and any new value will be determined when the plan is updated.

resourceChangeDetail_target :: Lens' ResourceChangeDetail (Maybe ResourceTargetDefinition)

Information about the resource attribute to be modified.