Network.AWS.ServiceCatalog.Types.StackInstance

Description

Documentation

data StackInstance

An AWS CloudFormation stack, in a specific account and region, that's part of a stack set operation. A stack instance is a reference to an attempted or actual stack in a given account within a given region. A stack instance can exist without a stack—for example, if the stack couldn't be created for some reason. A stack instance is associated with only one stack set. Each stack instance contains the ID of its associated stack set, as well as the ID of the actual stack and the stack status.

See: newStackInstance smart constructor.

Constructors

StackInstance'

Fields

stackInstanceStatus :: Maybe StackInstanceStatus

The status of the stack instance, in terms of its synchronization with its associated stack set.

INOPERABLE: A DeleteStackInstances operation has failed and left the stack in an unstable state. Stacks in this state are excluded from further UpdateStackSet operations. You might need to perform a DeleteStackInstances operation, with RetainStacks set to true, to delete the stack instance, and then delete the stack manually.

OUTDATED: The stack isn't currently up to date with the stack set because either the associated stack failed during a CreateStackSet or UpdateStackSet operation, or the stack was part of a CreateStackSet or UpdateStackSet operation that failed or was stopped before the stack was created or updated.

CURRENT: The stack is currently up to date with the stack set.

account :: Maybe Text

The name of the AWS account that the stack instance is associated with.

region :: Maybe Text

The name of the AWS region that the stack instance is associated with.

newStackInstance :: StackInstance

Create a value of StackInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackInstanceStatus:StackInstance', stackInstance_stackInstanceStatus - The status of the stack instance, in terms of its synchronization with its associated stack set.

INOPERABLE: A DeleteStackInstances operation has failed and left the stack in an unstable state. Stacks in this state are excluded from further UpdateStackSet operations. You might need to perform a DeleteStackInstances operation, with RetainStacks set to true, to delete the stack instance, and then delete the stack manually.

OUTDATED: The stack isn't currently up to date with the stack set because either the associated stack failed during a CreateStackSet or UpdateStackSet operation, or the stack was part of a CreateStackSet or UpdateStackSet operation that failed or was stopped before the stack was created or updated.

CURRENT: The stack is currently up to date with the stack set.

$sel:account:StackInstance', stackInstance_account - The name of the AWS account that the stack instance is associated with.

$sel:region:StackInstance', stackInstance_region - The name of the AWS region that the stack instance is associated with.

stackInstance_stackInstanceStatus :: Lens' StackInstance (Maybe StackInstanceStatus)

The status of the stack instance, in terms of its synchronization with its associated stack set.

INOPERABLE: A DeleteStackInstances operation has failed and left the stack in an unstable state. Stacks in this state are excluded from further UpdateStackSet operations. You might need to perform a DeleteStackInstances operation, with RetainStacks set to true, to delete the stack instance, and then delete the stack manually.

OUTDATED: The stack isn't currently up to date with the stack set because either the associated stack failed during a CreateStackSet or UpdateStackSet operation, or the stack was part of a CreateStackSet or UpdateStackSet operation that failed or was stopped before the stack was created or updated.

CURRENT: The stack is currently up to date with the stack set.

stackInstance_account :: Lens' StackInstance (Maybe Text)

The name of the AWS account that the stack instance is associated with.

stackInstance_region :: Lens' StackInstance (Maybe Text)

The name of the AWS region that the stack instance is associated with.