Network.AWS.ServiceCatalog.Types.UpdateProvisioningPreferences

Description

Documentation

data UpdateProvisioningPreferences

The user-defined preferences that will be applied when updating a provisioned product. Not all preferences are applicable to all provisioned product types.

See: newUpdateProvisioningPreferences smart constructor.

Constructors

UpdateProvisioningPreferences'

Fields

stackSetOperationType :: Maybe StackSetOperationType

Determines what action AWS Service Catalog performs to a stack set or a stack instance represented by the provisioned product. The default value is UPDATE if nothing is specified.

Applicable only to a CFN_STACKSET provisioned product type.

CREATE

Creates a new stack instance in the stack set represented by the provisioned product. In this case, only new stack instances are created based on accounts and regions; if new ProductId or ProvisioningArtifactID are passed, they will be ignored.

UPDATE

Updates the stack set represented by the provisioned product and also its stack instances.

DELETE

Deletes a stack instance in the stack set represented by the provisioned product.

stackSetAccounts :: Maybe [Text]

One or more AWS accounts that will have access to the provisioned product.

Applicable only to a CFN_STACKSET provisioned product type.

The AWS accounts specified should be within the list of accounts in the STACKSET constraint. To get the list of accounts in the STACKSET constraint, use the DescribeProvisioningParameters operation.

If no values are specified, the default value is all accounts from the STACKSET constraint.

stackSetFailureToleranceCount :: Maybe Natural

The number of accounts, per region, for which this operation can fail before AWS Service Catalog stops the operation in that region. If the operation is stopped in a region, AWS Service Catalog doesn't attempt the operation in any subsequent regions.

Applicable only to a CFN_STACKSET provisioned product type.

Conditional: You must specify either StackSetFailureToleranceCount or StackSetFailureTolerancePercentage, but not both.

The default value is 0 if no value is specified.

stackSetFailureTolerancePercentage :: Maybe Natural

The percentage of accounts, per region, for which this stack operation can fail before AWS Service Catalog stops the operation in that region. If the operation is stopped in a region, AWS Service Catalog doesn't attempt the operation in any subsequent regions.

When calculating the number of accounts based on the specified percentage, AWS Service Catalog rounds down to the next whole number.

Applicable only to a CFN_STACKSET provisioned product type.

Conditional: You must specify either StackSetFailureToleranceCount or StackSetFailureTolerancePercentage, but not both.

stackSetRegions :: Maybe [Text]

One or more AWS Regions where the provisioned product will be available.

Applicable only to a CFN_STACKSET provisioned product type.

The specified regions should be within the list of regions from the STACKSET constraint. To get the list of regions in the STACKSET constraint, use the DescribeProvisioningParameters operation.

If no values are specified, the default value is all regions from the STACKSET constraint.

stackSetMaxConcurrencyCount :: Maybe Natural

The maximum number of accounts in which to perform this operation at one time. This is dependent on the value of StackSetFailureToleranceCount. StackSetMaxConcurrentCount is at most one more than the StackSetFailureToleranceCount.

Note that this setting lets you specify the maximum for operations. For large deployments, under certain circumstances the actual number of accounts acted upon concurrently may be lower due to service throttling.

Applicable only to a CFN_STACKSET provisioned product type.

Conditional: You must specify either StackSetMaxConcurrentCount or StackSetMaxConcurrentPercentage, but not both.

stackSetMaxConcurrencyPercentage :: Maybe Natural

The maximum percentage of accounts in which to perform this operation at one time.

When calculating the number of accounts based on the specified percentage, AWS Service Catalog rounds down to the next whole number. This is true except in cases where rounding down would result is zero. In this case, AWS Service Catalog sets the number as 1 instead.

Note that this setting lets you specify the maximum for operations. For large deployments, under certain circumstances the actual number of accounts acted upon concurrently may be lower due to service throttling.

Applicable only to a CFN_STACKSET provisioned product type.

Conditional: You must specify either StackSetMaxConcurrentCount or StackSetMaxConcurrentPercentage, but not both.

newUpdateProvisioningPreferences :: UpdateProvisioningPreferences

Create a value of UpdateProvisioningPreferences with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackSetOperationType:UpdateProvisioningPreferences', updateProvisioningPreferences_stackSetOperationType - Determines what action AWS Service Catalog performs to a stack set or a stack instance represented by the provisioned product. The default value is UPDATE if nothing is specified.

Applicable only to a CFN_STACKSET provisioned product type.

CREATE

Creates a new stack instance in the stack set represented by the provisioned product. In this case, only new stack instances are created based on accounts and regions; if new ProductId or ProvisioningArtifactID are passed, they will be ignored.

UPDATE

Updates the stack set represented by the provisioned product and also its stack instances.

DELETE

Deletes a stack instance in the stack set represented by the provisioned product.

$sel:stackSetAccounts:UpdateProvisioningPreferences', updateProvisioningPreferences_stackSetAccounts - One or more AWS accounts that will have access to the provisioned product.

Applicable only to a CFN_STACKSET provisioned product type.

The AWS accounts specified should be within the list of accounts in the STACKSET constraint. To get the list of accounts in the STACKSET constraint, use the DescribeProvisioningParameters operation.

If no values are specified, the default value is all accounts from the STACKSET constraint.

$sel:stackSetFailureToleranceCount:UpdateProvisioningPreferences', updateProvisioningPreferences_stackSetFailureToleranceCount - The number of accounts, per region, for which this operation can fail before AWS Service Catalog stops the operation in that region. If the operation is stopped in a region, AWS Service Catalog doesn't attempt the operation in any subsequent regions.

Applicable only to a CFN_STACKSET provisioned product type.

Conditional: You must specify either StackSetFailureToleranceCount or StackSetFailureTolerancePercentage, but not both.

The default value is 0 if no value is specified.

$sel:stackSetFailureTolerancePercentage:UpdateProvisioningPreferences', updateProvisioningPreferences_stackSetFailureTolerancePercentage - The percentage of accounts, per region, for which this stack operation can fail before AWS Service Catalog stops the operation in that region. If the operation is stopped in a region, AWS Service Catalog doesn't attempt the operation in any subsequent regions.

When calculating the number of accounts based on the specified percentage, AWS Service Catalog rounds down to the next whole number.

Applicable only to a CFN_STACKSET provisioned product type.

Conditional: You must specify either StackSetFailureToleranceCount or StackSetFailureTolerancePercentage, but not both.

$sel:stackSetRegions:UpdateProvisioningPreferences', updateProvisioningPreferences_stackSetRegions - One or more AWS Regions where the provisioned product will be available.

Applicable only to a CFN_STACKSET provisioned product type.

The specified regions should be within the list of regions from the STACKSET constraint. To get the list of regions in the STACKSET constraint, use the DescribeProvisioningParameters operation.

If no values are specified, the default value is all regions from the STACKSET constraint.

$sel:stackSetMaxConcurrencyCount:UpdateProvisioningPreferences', updateProvisioningPreferences_stackSetMaxConcurrencyCount - The maximum number of accounts in which to perform this operation at one time. This is dependent on the value of StackSetFailureToleranceCount. StackSetMaxConcurrentCount is at most one more than the StackSetFailureToleranceCount.

Note that this setting lets you specify the maximum for operations. For large deployments, under certain circumstances the actual number of accounts acted upon concurrently may be lower due to service throttling.

Applicable only to a CFN_STACKSET provisioned product type.

Conditional: You must specify either StackSetMaxConcurrentCount or StackSetMaxConcurrentPercentage, but not both.

$sel:stackSetMaxConcurrencyPercentage:UpdateProvisioningPreferences', updateProvisioningPreferences_stackSetMaxConcurrencyPercentage - The maximum percentage of accounts in which to perform this operation at one time.

When calculating the number of accounts based on the specified percentage, AWS Service Catalog rounds down to the next whole number. This is true except in cases where rounding down would result is zero. In this case, AWS Service Catalog sets the number as 1 instead.

Note that this setting lets you specify the maximum for operations. For large deployments, under certain circumstances the actual number of accounts acted upon concurrently may be lower due to service throttling.

Applicable only to a CFN_STACKSET provisioned product type.

Conditional: You must specify either StackSetMaxConcurrentCount or StackSetMaxConcurrentPercentage, but not both.

updateProvisioningPreferences_stackSetOperationType :: Lens' UpdateProvisioningPreferences (Maybe StackSetOperationType)

Determines what action AWS Service Catalog performs to a stack set or a stack instance represented by the provisioned product. The default value is UPDATE if nothing is specified.

Applicable only to a CFN_STACKSET provisioned product type.

CREATE

Creates a new stack instance in the stack set represented by the provisioned product. In this case, only new stack instances are created based on accounts and regions; if new ProductId or ProvisioningArtifactID are passed, they will be ignored.

UPDATE

Updates the stack set represented by the provisioned product and also its stack instances.

DELETE

Deletes a stack instance in the stack set represented by the provisioned product.

updateProvisioningPreferences_stackSetAccounts :: Lens' UpdateProvisioningPreferences (Maybe [Text])

One or more AWS accounts that will have access to the provisioned product.

Applicable only to a CFN_STACKSET provisioned product type.

The AWS accounts specified should be within the list of accounts in the STACKSET constraint. To get the list of accounts in the STACKSET constraint, use the DescribeProvisioningParameters operation.

If no values are specified, the default value is all accounts from the STACKSET constraint.

updateProvisioningPreferences_stackSetFailureToleranceCount :: Lens' UpdateProvisioningPreferences (Maybe Natural)

The number of accounts, per region, for which this operation can fail before AWS Service Catalog stops the operation in that region. If the operation is stopped in a region, AWS Service Catalog doesn't attempt the operation in any subsequent regions.

Applicable only to a CFN_STACKSET provisioned product type.

Conditional: You must specify either StackSetFailureToleranceCount or StackSetFailureTolerancePercentage, but not both.

The default value is 0 if no value is specified.

updateProvisioningPreferences_stackSetFailureTolerancePercentage :: Lens' UpdateProvisioningPreferences (Maybe Natural)

The percentage of accounts, per region, for which this stack operation can fail before AWS Service Catalog stops the operation in that region. If the operation is stopped in a region, AWS Service Catalog doesn't attempt the operation in any subsequent regions.

When calculating the number of accounts based on the specified percentage, AWS Service Catalog rounds down to the next whole number.

Applicable only to a CFN_STACKSET provisioned product type.

Conditional: You must specify either StackSetFailureToleranceCount or StackSetFailureTolerancePercentage, but not both.

updateProvisioningPreferences_stackSetRegions :: Lens' UpdateProvisioningPreferences (Maybe [Text])

One or more AWS Regions where the provisioned product will be available.

Applicable only to a CFN_STACKSET provisioned product type.

The specified regions should be within the list of regions from the STACKSET constraint. To get the list of regions in the STACKSET constraint, use the DescribeProvisioningParameters operation.

If no values are specified, the default value is all regions from the STACKSET constraint.

updateProvisioningPreferences_stackSetMaxConcurrencyCount :: Lens' UpdateProvisioningPreferences (Maybe Natural)

The maximum number of accounts in which to perform this operation at one time. This is dependent on the value of StackSetFailureToleranceCount. StackSetMaxConcurrentCount is at most one more than the StackSetFailureToleranceCount.

Note that this setting lets you specify the maximum for operations. For large deployments, under certain circumstances the actual number of accounts acted upon concurrently may be lower due to service throttling.

Applicable only to a CFN_STACKSET provisioned product type.

Conditional: You must specify either StackSetMaxConcurrentCount or StackSetMaxConcurrentPercentage, but not both.

updateProvisioningPreferences_stackSetMaxConcurrencyPercentage :: Lens' UpdateProvisioningPreferences (Maybe Natural)

The maximum percentage of accounts in which to perform this operation at one time.

When calculating the number of accounts based on the specified percentage, AWS Service Catalog rounds down to the next whole number. This is true except in cases where rounding down would result is zero. In this case, AWS Service Catalog sets the number as 1 instead.

Note that this setting lets you specify the maximum for operations. For large deployments, under certain circumstances the actual number of accounts acted upon concurrently may be lower due to service throttling.

Applicable only to a CFN_STACKSET provisioned product type.

Conditional: You must specify either StackSetMaxConcurrentCount or StackSetMaxConcurrentPercentage, but not both.