Network.AWS.ServiceCatalog.UpdateConstraint

Description

Updates the specified constraint.

Creating a Request

data UpdateConstraint

See: newUpdateConstraint smart constructor.

Constructors

UpdateConstraint'

Fields

description :: Maybe Text

The updated description of the constraint.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

parameters :: Maybe Text

The constraint parameters, in JSON format. The syntax depends on the constraint type as follows:

LAUNCH

You are required to specify either the RoleArn or the LocalRoleName but can't use both.
Specify the RoleArn property as follows:

{"RoleArn" : "arn:aws:iam::123456789012:role/LaunchRole"}

Specify the LocalRoleName property as follows:

{"LocalRoleName": "SCBasicLaunchRole"}

If you specify the LocalRoleName property, when an account uses the launch constraint, the IAM role with that name in the account will be used. This allows launch-role constraints to be account-agnostic so the administrator can create fewer resources per shared account.

The given role name must exist in the account used to create the launch constraint and the account of the user who launches a product with this launch constraint.

You cannot have both a LAUNCH and a STACKSET constraint.

You also cannot have more than one LAUNCH constraint on a product and portfolio.

NOTIFICATION

Specify the NotificationArns property as follows:
{"NotificationArns" : ["arn:aws:sns:us-east-1:123456789012:Topic"]}

RESOURCE_UPDATE

Specify the TagUpdatesOnProvisionedProduct property as follows:
{"Version":"2.0","Properties":{"TagUpdateOnProvisionedProduct":"String"}}

The TagUpdatesOnProvisionedProduct property accepts a string value of ALLOWED or NOT_ALLOWED.

STACKSET

Specify the Parameters property as follows:
{"Version": "String", "Properties": {"AccountList": [ "String" ], "RegionList": [ "String" ], "AdminRole": "String", "ExecutionRole": "String"}}

You cannot have both a LAUNCH and a STACKSET constraint.

You also cannot have more than one STACKSET constraint on a product and portfolio.

Products with a STACKSET constraint will launch an AWS CloudFormation stack set.

TEMPLATE

Specify the Rules property. For more information, see Template Constraint Rules.

id :: Text

The identifier of the constraint.

newUpdateConstraint

Arguments

:: Text
$sel:id:UpdateConstraint'

-> UpdateConstraint

Create a value of UpdateConstraint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:UpdateConstraint', updateConstraint_description - The updated description of the constraint.

$sel:acceptLanguage:UpdateConstraint', updateConstraint_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:parameters:UpdateConstraint', updateConstraint_parameters - The constraint parameters, in JSON format. The syntax depends on the constraint type as follows:

LAUNCH

You are required to specify either the RoleArn or the LocalRoleName but can't use both.
Specify the RoleArn property as follows:

{"RoleArn" : "arn:aws:iam::123456789012:role/LaunchRole"}

Specify the LocalRoleName property as follows:

{"LocalRoleName": "SCBasicLaunchRole"}

If you specify the LocalRoleName property, when an account uses the launch constraint, the IAM role with that name in the account will be used. This allows launch-role constraints to be account-agnostic so the administrator can create fewer resources per shared account.

The given role name must exist in the account used to create the launch constraint and the account of the user who launches a product with this launch constraint.

You cannot have both a LAUNCH and a STACKSET constraint.

You also cannot have more than one LAUNCH constraint on a product and portfolio.

NOTIFICATION

Specify the NotificationArns property as follows:
{"NotificationArns" : ["arn:aws:sns:us-east-1:123456789012:Topic"]}

RESOURCE_UPDATE

Specify the TagUpdatesOnProvisionedProduct property as follows:
{"Version":"2.0","Properties":{"TagUpdateOnProvisionedProduct":"String"}}

The TagUpdatesOnProvisionedProduct property accepts a string value of ALLOWED or NOT_ALLOWED.

STACKSET

Specify the Parameters property as follows:
{"Version": "String", "Properties": {"AccountList": [ "String" ], "RegionList": [ "String" ], "AdminRole": "String", "ExecutionRole": "String"}}

You cannot have both a LAUNCH and a STACKSET constraint.

You also cannot have more than one STACKSET constraint on a product and portfolio.

Products with a STACKSET constraint will launch an AWS CloudFormation stack set.

TEMPLATE

Specify the Rules property. For more information, see Template Constraint Rules.

$sel:id:UpdateConstraint', updateConstraint_id - The identifier of the constraint.

Request Lenses

updateConstraint_description :: Lens' UpdateConstraint (Maybe Text)

The updated description of the constraint.

updateConstraint_acceptLanguage :: Lens' UpdateConstraint (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

updateConstraint_parameters :: Lens' UpdateConstraint (Maybe Text)

The constraint parameters, in JSON format. The syntax depends on the constraint type as follows:

LAUNCH

You are required to specify either the RoleArn or the LocalRoleName but can't use both.
Specify the RoleArn property as follows:

{"RoleArn" : "arn:aws:iam::123456789012:role/LaunchRole"}

Specify the LocalRoleName property as follows:

{"LocalRoleName": "SCBasicLaunchRole"}

If you specify the LocalRoleName property, when an account uses the launch constraint, the IAM role with that name in the account will be used. This allows launch-role constraints to be account-agnostic so the administrator can create fewer resources per shared account.

The given role name must exist in the account used to create the launch constraint and the account of the user who launches a product with this launch constraint.

You cannot have both a LAUNCH and a STACKSET constraint.

You also cannot have more than one LAUNCH constraint on a product and portfolio.

NOTIFICATION

Specify the NotificationArns property as follows:
{"NotificationArns" : ["arn:aws:sns:us-east-1:123456789012:Topic"]}

RESOURCE_UPDATE

Specify the TagUpdatesOnProvisionedProduct property as follows:
{"Version":"2.0","Properties":{"TagUpdateOnProvisionedProduct":"String"}}

The TagUpdatesOnProvisionedProduct property accepts a string value of ALLOWED or NOT_ALLOWED.

STACKSET

Specify the Parameters property as follows:
{"Version": "String", "Properties": {"AccountList": [ "String" ], "RegionList": [ "String" ], "AdminRole": "String", "ExecutionRole": "String"}}

You cannot have both a LAUNCH and a STACKSET constraint.

You also cannot have more than one STACKSET constraint on a product and portfolio.

Products with a STACKSET constraint will launch an AWS CloudFormation stack set.

TEMPLATE

Specify the Rules property. For more information, see Template Constraint Rules.

updateConstraint_id :: Lens' UpdateConstraint Text

The identifier of the constraint.

Destructuring the Response

data UpdateConstraintResponse

See: newUpdateConstraintResponse smart constructor.

Constructors

UpdateConstraintResponse'

Fields

constraintParameters :: Maybe Text

The constraint parameters.

constraintDetail :: Maybe ConstraintDetail

Information about the constraint.

status :: Maybe RequestStatus

The status of the current request.

httpStatus :: Int

The response's http status code.

newUpdateConstraintResponse

Arguments

:: Int
$sel:httpStatus:UpdateConstraintResponse'

-> UpdateConstraintResponse

Create a value of UpdateConstraintResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:constraintParameters:UpdateConstraintResponse', updateConstraintResponse_constraintParameters - The constraint parameters.

$sel:constraintDetail:UpdateConstraintResponse', updateConstraintResponse_constraintDetail - Information about the constraint.

$sel:status:UpdateConstraintResponse', updateConstraintResponse_status - The status of the current request.

$sel:httpStatus:UpdateConstraintResponse', updateConstraintResponse_httpStatus - The response's http status code.

Response Lenses

updateConstraintResponse_constraintParameters :: Lens' UpdateConstraintResponse (Maybe Text)

The constraint parameters.

updateConstraintResponse_constraintDetail :: Lens' UpdateConstraintResponse (Maybe ConstraintDetail)

Information about the constraint.

updateConstraintResponse_status :: Lens' UpdateConstraintResponse (Maybe RequestStatus)

The status of the current request.

updateConstraintResponse_httpStatus :: Lens' UpdateConstraintResponse Int

The response's http status code.