Network.AWS.ServiceCatalog.UpdateProduct

Description

Updates the specified product.

Creating a Request

data UpdateProduct

See: newUpdateProduct smart constructor.

Constructors

UpdateProduct'

Fields

removeTags :: Maybe [Text]

The tags to remove from the product.

owner :: Maybe Text

The updated owner of the product.

supportUrl :: Maybe Text

The updated support URL for the product.

distributor :: Maybe Text

The updated distributor of the product.

name :: Maybe Text

The updated product name.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

addTags :: Maybe [Tag]

The tags to add to the product.

supportEmail :: Maybe Text

The updated support email for the product.

description :: Maybe Text

The updated description of the product.

supportDescription :: Maybe Text

The updated support description for the product.

id :: Text

The product identifier.

newUpdateProduct

Arguments

:: Text
$sel:id:UpdateProduct'

-> UpdateProduct

Create a value of UpdateProduct with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:removeTags:UpdateProduct', updateProduct_removeTags - The tags to remove from the product.

$sel:owner:UpdateProduct', updateProduct_owner - The updated owner of the product.

$sel:supportUrl:UpdateProduct', updateProduct_supportUrl - The updated support URL for the product.

$sel:distributor:UpdateProduct', updateProduct_distributor - The updated distributor of the product.

$sel:name:UpdateProduct', updateProduct_name - The updated product name.

$sel:acceptLanguage:UpdateProduct', updateProduct_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:addTags:UpdateProduct', updateProduct_addTags - The tags to add to the product.

$sel:supportEmail:UpdateProduct', updateProduct_supportEmail - The updated support email for the product.

$sel:description:UpdateProduct', updateProduct_description - The updated description of the product.

$sel:supportDescription:UpdateProduct', updateProduct_supportDescription - The updated support description for the product.

$sel:id:UpdateProduct', updateProduct_id - The product identifier.

Request Lenses

updateProduct_removeTags :: Lens' UpdateProduct (Maybe [Text])

The tags to remove from the product.

updateProduct_owner :: Lens' UpdateProduct (Maybe Text)

The updated owner of the product.

updateProduct_supportUrl :: Lens' UpdateProduct (Maybe Text)

The updated support URL for the product.

updateProduct_distributor :: Lens' UpdateProduct (Maybe Text)

The updated distributor of the product.

updateProduct_name :: Lens' UpdateProduct (Maybe Text)

The updated product name.

updateProduct_acceptLanguage :: Lens' UpdateProduct (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

updateProduct_addTags :: Lens' UpdateProduct (Maybe [Tag])

The tags to add to the product.

updateProduct_supportEmail :: Lens' UpdateProduct (Maybe Text)

The updated support email for the product.

updateProduct_description :: Lens' UpdateProduct (Maybe Text)

The updated description of the product.

updateProduct_supportDescription :: Lens' UpdateProduct (Maybe Text)

The updated support description for the product.

updateProduct_id :: Lens' UpdateProduct Text

The product identifier.

Destructuring the Response

data UpdateProductResponse

See: newUpdateProductResponse smart constructor.

Constructors

UpdateProductResponse'

Fields

productViewDetail :: Maybe ProductViewDetail

Information about the product view.

tags :: Maybe [Tag]

Information about the tags associated with the product.

httpStatus :: Int

The response's http status code.

newUpdateProductResponse

Arguments

:: Int
$sel:httpStatus:UpdateProductResponse'

-> UpdateProductResponse

Create a value of UpdateProductResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:productViewDetail:UpdateProductResponse', updateProductResponse_productViewDetail - Information about the product view.

$sel:tags:UpdateProductResponse', updateProductResponse_tags - Information about the tags associated with the product.

$sel:httpStatus:UpdateProductResponse', updateProductResponse_httpStatus - The response's http status code.

Response Lenses

updateProductResponse_productViewDetail :: Lens' UpdateProductResponse (Maybe ProductViewDetail)

Information about the product view.

updateProductResponse_tags :: Lens' UpdateProductResponse (Maybe [Tag])

Information about the tags associated with the product.

updateProductResponse_httpStatus :: Lens' UpdateProductResponse Int

The response's http status code.