Network.AWS.ServiceCatalog.UpdateProvisionedProduct

Description

Requests updates to the configuration of the specified provisioned product.

If there are tags associated with the object, they cannot be updated or added. Depending on the specific updates requested, this operation can update with no interruption, with some interruption, or replace the provisioned product entirely.

You can check the status of this request using DescribeRecord.

Creating a Request

data UpdateProvisionedProduct

See: newUpdateProvisionedProduct smart constructor.

Constructors

UpdateProvisionedProduct'

Fields

productName :: Maybe Text

The name of the product. You must provide the name or ID, but not both.

provisionedProductName :: Maybe Text

The name of the provisioned product. You cannot specify both ProvisionedProductName and ProvisionedProductId.

provisioningArtifactId :: Maybe Text

The identifier of the provisioning artifact.

provisioningArtifactName :: Maybe Text

The name of the provisioning artifact. You must provide the name or ID, but not both.

pathName :: Maybe Text

The name of the path. You must provide the name or ID, but not both.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

pathId :: Maybe Text

The path identifier. This value is optional if the product has a default path, and required if the product has more than one path. You must provide the name or ID, but not both.

provisioningParameters :: Maybe [UpdateProvisioningParameter]

The new parameters.

provisionedProductId :: Maybe Text

The identifier of the provisioned product. You must provide the name or ID, but not both.

productId :: Maybe Text

The identifier of the product. You must provide the name or ID, but not both.

tags :: Maybe [Tag]

One or more tags. Requires the product to have RESOURCE_UPDATE constraint with TagUpdatesOnProvisionedProduct set to ALLOWED to allow tag updates.

provisioningPreferences :: Maybe UpdateProvisioningPreferences

An object that contains information about the provisioning preferences for a stack set.

updateToken :: Text

The idempotency token that uniquely identifies the provisioning update request.

newUpdateProvisionedProduct

Arguments

:: Text
$sel:updateToken:UpdateProvisionedProduct'

-> UpdateProvisionedProduct

Create a value of UpdateProvisionedProduct with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:productName:UpdateProvisionedProduct', updateProvisionedProduct_productName - The name of the product. You must provide the name or ID, but not both.

$sel:provisionedProductName:UpdateProvisionedProduct', updateProvisionedProduct_provisionedProductName - The name of the provisioned product. You cannot specify both ProvisionedProductName and ProvisionedProductId.

$sel:provisioningArtifactId:UpdateProvisionedProduct', updateProvisionedProduct_provisioningArtifactId - The identifier of the provisioning artifact.

$sel:provisioningArtifactName:UpdateProvisionedProduct', updateProvisionedProduct_provisioningArtifactName - The name of the provisioning artifact. You must provide the name or ID, but not both.

$sel:pathName:UpdateProvisionedProduct', updateProvisionedProduct_pathName - The name of the path. You must provide the name or ID, but not both.

$sel:acceptLanguage:UpdateProvisionedProduct', updateProvisionedProduct_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:pathId:UpdateProvisionedProduct', updateProvisionedProduct_pathId - The path identifier. This value is optional if the product has a default path, and required if the product has more than one path. You must provide the name or ID, but not both.

$sel:provisioningParameters:UpdateProvisionedProduct', updateProvisionedProduct_provisioningParameters - The new parameters.

$sel:provisionedProductId:UpdateProvisionedProduct', updateProvisionedProduct_provisionedProductId - The identifier of the provisioned product. You must provide the name or ID, but not both.

$sel:productId:UpdateProvisionedProduct', updateProvisionedProduct_productId - The identifier of the product. You must provide the name or ID, but not both.

$sel:tags:UpdateProvisionedProduct', updateProvisionedProduct_tags - One or more tags. Requires the product to have RESOURCE_UPDATE constraint with TagUpdatesOnProvisionedProduct set to ALLOWED to allow tag updates.

$sel:provisioningPreferences:UpdateProvisionedProduct', updateProvisionedProduct_provisioningPreferences - An object that contains information about the provisioning preferences for a stack set.

$sel:updateToken:UpdateProvisionedProduct', updateProvisionedProduct_updateToken - The idempotency token that uniquely identifies the provisioning update request.

Request Lenses

updateProvisionedProduct_productName :: Lens' UpdateProvisionedProduct (Maybe Text)

The name of the product. You must provide the name or ID, but not both.

updateProvisionedProduct_provisionedProductName :: Lens' UpdateProvisionedProduct (Maybe Text)

The name of the provisioned product. You cannot specify both ProvisionedProductName and ProvisionedProductId.

updateProvisionedProduct_provisioningArtifactId :: Lens' UpdateProvisionedProduct (Maybe Text)

The identifier of the provisioning artifact.

updateProvisionedProduct_provisioningArtifactName :: Lens' UpdateProvisionedProduct (Maybe Text)

The name of the provisioning artifact. You must provide the name or ID, but not both.

updateProvisionedProduct_pathName :: Lens' UpdateProvisionedProduct (Maybe Text)

The name of the path. You must provide the name or ID, but not both.

updateProvisionedProduct_acceptLanguage :: Lens' UpdateProvisionedProduct (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

updateProvisionedProduct_pathId :: Lens' UpdateProvisionedProduct (Maybe Text)

The path identifier. This value is optional if the product has a default path, and required if the product has more than one path. You must provide the name or ID, but not both.

updateProvisionedProduct_provisioningParameters :: Lens' UpdateProvisionedProduct (Maybe [UpdateProvisioningParameter])

The new parameters.

updateProvisionedProduct_provisionedProductId :: Lens' UpdateProvisionedProduct (Maybe Text)

The identifier of the provisioned product. You must provide the name or ID, but not both.

updateProvisionedProduct_productId :: Lens' UpdateProvisionedProduct (Maybe Text)

The identifier of the product. You must provide the name or ID, but not both.

updateProvisionedProduct_tags :: Lens' UpdateProvisionedProduct (Maybe [Tag])

One or more tags. Requires the product to have RESOURCE_UPDATE constraint with TagUpdatesOnProvisionedProduct set to ALLOWED to allow tag updates.

updateProvisionedProduct_provisioningPreferences :: Lens' UpdateProvisionedProduct (Maybe UpdateProvisioningPreferences)

An object that contains information about the provisioning preferences for a stack set.

updateProvisionedProduct_updateToken :: Lens' UpdateProvisionedProduct Text

The idempotency token that uniquely identifies the provisioning update request.

Destructuring the Response

data UpdateProvisionedProductResponse

See: newUpdateProvisionedProductResponse smart constructor.

Constructors

UpdateProvisionedProductResponse'

Fields

recordDetail :: Maybe RecordDetail

Information about the result of the request.

httpStatus :: Int

The response's http status code.

newUpdateProvisionedProductResponse

Arguments

:: Int
$sel:httpStatus:UpdateProvisionedProductResponse'

-> UpdateProvisionedProductResponse

Create a value of UpdateProvisionedProductResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recordDetail:UpdateProvisionedProductResponse', updateProvisionedProductResponse_recordDetail - Information about the result of the request.

$sel:httpStatus:UpdateProvisionedProductResponse', updateProvisionedProductResponse_httpStatus - The response's http status code.

Response Lenses

updateProvisionedProductResponse_recordDetail :: Lens' UpdateProvisionedProductResponse (Maybe RecordDetail)

Information about the result of the request.

updateProvisionedProductResponse_httpStatus :: Lens' UpdateProvisionedProductResponse Int

The response's http status code.