Network.AWS.ServiceCatalogAppRegistry.GetAttributeGroup

Description

Retrieves an attribute group, either by its name or its ID. The attribute group can be specified either by its unique ID or by its name.

Creating a Request

data GetAttributeGroup

See: newGetAttributeGroup smart constructor.

Constructors

GetAttributeGroup'

Fields

attributeGroup :: Text

The name or ID of the attribute group that holds the attributes to describe the application.

newGetAttributeGroup

Arguments

:: Text
$sel:attributeGroup:GetAttributeGroup'

-> GetAttributeGroup

Create a value of GetAttributeGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeGroup:GetAttributeGroup', getAttributeGroup_attributeGroup - The name or ID of the attribute group that holds the attributes to describe the application.

Request Lenses

getAttributeGroup_attributeGroup :: Lens' GetAttributeGroup Text

The name or ID of the attribute group that holds the attributes to describe the application.

Destructuring the Response

data GetAttributeGroupResponse

See: newGetAttributeGroupResponse smart constructor.

Constructors

GetAttributeGroupResponse'

Fields

creationTime :: Maybe POSIX

The ISO-8601 formatted timestamp of the moment the attribute group was created.

arn :: Maybe Text

The Amazon resource name (ARN) that specifies the attribute group across services.

name :: Maybe Text

The name of the attribute group.

attributes :: Maybe Text

A JSON string in the form of nested key-value pairs that represent the attributes in the group and describes an application and its components.

id :: Maybe Text

The identifier of the attribute group.

lastUpdateTime :: Maybe POSIX

The ISO-8601 formatted timestamp of the moment the attribute group was last updated. This time is the same as the creationTime for a newly created attribute group.

description :: Maybe Text

The description of the attribute group that the user provides.

tags :: Maybe (HashMap Text Text)

Key-value pairs associated with the attribute group.

httpStatus :: Int

The response's http status code.

newGetAttributeGroupResponse

Arguments

:: Int
$sel:httpStatus:GetAttributeGroupResponse'

-> GetAttributeGroupResponse

Create a value of GetAttributeGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:GetAttributeGroupResponse', getAttributeGroupResponse_creationTime - The ISO-8601 formatted timestamp of the moment the attribute group was created.

$sel:arn:GetAttributeGroupResponse', getAttributeGroupResponse_arn - The Amazon resource name (ARN) that specifies the attribute group across services.

$sel:name:GetAttributeGroupResponse', getAttributeGroupResponse_name - The name of the attribute group.

$sel:attributes:GetAttributeGroupResponse', getAttributeGroupResponse_attributes - A JSON string in the form of nested key-value pairs that represent the attributes in the group and describes an application and its components.

$sel:id:GetAttributeGroupResponse', getAttributeGroupResponse_id - The identifier of the attribute group.

$sel:lastUpdateTime:GetAttributeGroupResponse', getAttributeGroupResponse_lastUpdateTime - The ISO-8601 formatted timestamp of the moment the attribute group was last updated. This time is the same as the creationTime for a newly created attribute group.

$sel:description:GetAttributeGroupResponse', getAttributeGroupResponse_description - The description of the attribute group that the user provides.

$sel:tags:GetAttributeGroupResponse', getAttributeGroupResponse_tags - Key-value pairs associated with the attribute group.

$sel:httpStatus:GetAttributeGroupResponse', getAttributeGroupResponse_httpStatus - The response's http status code.

Response Lenses

getAttributeGroupResponse_creationTime :: Lens' GetAttributeGroupResponse (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment the attribute group was created.

getAttributeGroupResponse_arn :: Lens' GetAttributeGroupResponse (Maybe Text)

The Amazon resource name (ARN) that specifies the attribute group across services.

getAttributeGroupResponse_name :: Lens' GetAttributeGroupResponse (Maybe Text)

The name of the attribute group.

getAttributeGroupResponse_attributes :: Lens' GetAttributeGroupResponse (Maybe Text)

A JSON string in the form of nested key-value pairs that represent the attributes in the group and describes an application and its components.

getAttributeGroupResponse_id :: Lens' GetAttributeGroupResponse (Maybe Text)

The identifier of the attribute group.

getAttributeGroupResponse_lastUpdateTime :: Lens' GetAttributeGroupResponse (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment the attribute group was last updated. This time is the same as the creationTime for a newly created attribute group.

getAttributeGroupResponse_description :: Lens' GetAttributeGroupResponse (Maybe Text)

The description of the attribute group that the user provides.

getAttributeGroupResponse_tags :: Lens' GetAttributeGroupResponse (Maybe (HashMap Text Text))

Key-value pairs associated with the attribute group.

getAttributeGroupResponse_httpStatus :: Lens' GetAttributeGroupResponse Int

The response's http status code.