Network.AWS.ServiceCatalogAppRegistry.ListAssociatedResources

Description

Lists all resources that are associated with specified application. Results are paginated.

This operation returns paginated results.

Creating a Request

data ListAssociatedResources

See: newListAssociatedResources smart constructor.

Constructors

ListAssociatedResources'

Fields

nextToken :: Maybe Text

The token to use to get the next page of results after a previous API call.

maxResults :: Maybe Natural

The upper bound of the number of results to return (cannot exceed 25). If this parameter is omitted, it defaults to 25. This value is optional.

application :: Text

The name or ID of the application.

newListAssociatedResources

Arguments

:: Text
$sel:application:ListAssociatedResources'

-> ListAssociatedResources

Create a value of ListAssociatedResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssociatedResources', listAssociatedResources_nextToken - The token to use to get the next page of results after a previous API call.

$sel:maxResults:ListAssociatedResources', listAssociatedResources_maxResults - The upper bound of the number of results to return (cannot exceed 25). If this parameter is omitted, it defaults to 25. This value is optional.

$sel:application:ListAssociatedResources', listAssociatedResources_application - The name or ID of the application.

Request Lenses

listAssociatedResources_nextToken :: Lens' ListAssociatedResources (Maybe Text)

The token to use to get the next page of results after a previous API call.

listAssociatedResources_maxResults :: Lens' ListAssociatedResources (Maybe Natural)

The upper bound of the number of results to return (cannot exceed 25). If this parameter is omitted, it defaults to 25. This value is optional.

listAssociatedResources_application :: Lens' ListAssociatedResources Text

The name or ID of the application.

Destructuring the Response

data ListAssociatedResourcesResponse

See: newListAssociatedResourcesResponse smart constructor.

Constructors

ListAssociatedResourcesResponse'

Fields

resources :: Maybe [ResourceInfo]

Information about the resources.

nextToken :: Maybe Text

The token to use to get the next page of results after a previous API call.

httpStatus :: Int

The response's http status code.

newListAssociatedResourcesResponse

Arguments

:: Int
$sel:httpStatus:ListAssociatedResourcesResponse'

-> ListAssociatedResourcesResponse

Create a value of ListAssociatedResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resources:ListAssociatedResourcesResponse', listAssociatedResourcesResponse_resources - Information about the resources.

$sel:nextToken:ListAssociatedResources', listAssociatedResourcesResponse_nextToken - The token to use to get the next page of results after a previous API call.

$sel:httpStatus:ListAssociatedResourcesResponse', listAssociatedResourcesResponse_httpStatus - The response's http status code.

Response Lenses

listAssociatedResourcesResponse_resources :: Lens' ListAssociatedResourcesResponse (Maybe [ResourceInfo])

Information about the resources.

listAssociatedResourcesResponse_nextToken :: Lens' ListAssociatedResourcesResponse (Maybe Text)

The token to use to get the next page of results after a previous API call.

listAssociatedResourcesResponse_httpStatus :: Lens' ListAssociatedResourcesResponse Int

The response's http status code.