Network.AWS.ServiceCatalogAppRegistry.Types.Application

Description

Documentation

data Application

Represents a Amazon Web Services Service Catalog AppRegistry application that is the top-level node in a hierarchy of related cloud resource abstractions.

See: newApplication smart constructor.

Constructors

Application'

Fields

creationTime :: Maybe POSIX

The ISO-8601 formatted timestamp of the moment when the application was created.

arn :: Maybe Text

The Amazon resource name (ARN) that specifies the application across services.

name :: Maybe Text

The name of the application. The name must be unique in the region in which you are creating the application.

id :: Maybe Text

The identifier of the application.

lastUpdateTime :: Maybe POSIX

The ISO-8601 formatted timestamp of the moment when the application was last updated.

description :: Maybe Text

The description of the application.

tags :: Maybe (HashMap Text Text)

Key-value pairs you can use to associate with the application.

newApplication :: Application

Create a value of Application with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Application', application_creationTime - The ISO-8601 formatted timestamp of the moment when the application was created.

$sel:arn:Application', application_arn - The Amazon resource name (ARN) that specifies the application across services.

$sel:name:Application', application_name - The name of the application. The name must be unique in the region in which you are creating the application.

$sel:id:Application', application_id - The identifier of the application.

$sel:lastUpdateTime:Application', application_lastUpdateTime - The ISO-8601 formatted timestamp of the moment when the application was last updated.

$sel:description:Application', application_description - The description of the application.

$sel:tags:Application', application_tags - Key-value pairs you can use to associate with the application.

application_creationTime :: Lens' Application (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment when the application was created.

application_arn :: Lens' Application (Maybe Text)

The Amazon resource name (ARN) that specifies the application across services.

application_name :: Lens' Application (Maybe Text)

The name of the application. The name must be unique in the region in which you are creating the application.

application_id :: Lens' Application (Maybe Text)

The identifier of the application.

application_lastUpdateTime :: Lens' Application (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment when the application was last updated.

application_description :: Lens' Application (Maybe Text)

The description of the application.

application_tags :: Lens' Application (Maybe (HashMap Text Text))

Key-value pairs you can use to associate with the application.