Network.AWS.ServiceCatalogAppRegistry.UpdateAttributeGroup

Description

Updates an existing attribute group with new details.

Creating a Request

data UpdateAttributeGroup

See: newUpdateAttributeGroup smart constructor.

Constructors

UpdateAttributeGroup'

Fields

name :: Maybe Text

The new name of the attribute group. The name must be unique in the region in which you are updating the attribute group.

attributes :: Maybe Text

A JSON string in the form of nested key-value pairs that represent the attributes in the group and describes an application and its components.

description :: Maybe Text

The description of the attribute group that the user provides.

attributeGroup :: Text

The name or ID of the attribute group that holds the attributes to describe the application.

newUpdateAttributeGroup

Arguments

:: Text
$sel:attributeGroup:UpdateAttributeGroup'

-> UpdateAttributeGroup

Create a value of UpdateAttributeGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateAttributeGroup', updateAttributeGroup_name - The new name of the attribute group. The name must be unique in the region in which you are updating the attribute group.

$sel:attributes:UpdateAttributeGroup', updateAttributeGroup_attributes - A JSON string in the form of nested key-value pairs that represent the attributes in the group and describes an application and its components.

$sel:description:UpdateAttributeGroup', updateAttributeGroup_description - The description of the attribute group that the user provides.

$sel:attributeGroup:UpdateAttributeGroup', updateAttributeGroup_attributeGroup - The name or ID of the attribute group that holds the attributes to describe the application.

Request Lenses

updateAttributeGroup_name :: Lens' UpdateAttributeGroup (Maybe Text)

The new name of the attribute group. The name must be unique in the region in which you are updating the attribute group.

updateAttributeGroup_attributes :: Lens' UpdateAttributeGroup (Maybe Text)

A JSON string in the form of nested key-value pairs that represent the attributes in the group and describes an application and its components.

updateAttributeGroup_description :: Lens' UpdateAttributeGroup (Maybe Text)

The description of the attribute group that the user provides.

updateAttributeGroup_attributeGroup :: Lens' UpdateAttributeGroup Text

The name or ID of the attribute group that holds the attributes to describe the application.

Destructuring the Response

data UpdateAttributeGroupResponse

See: newUpdateAttributeGroupResponse smart constructor.

Constructors

UpdateAttributeGroupResponse'

Fields

attributeGroup :: Maybe AttributeGroup

The updated information of the attribute group.

httpStatus :: Int

The response's http status code.

newUpdateAttributeGroupResponse

Arguments

:: Int
$sel:httpStatus:UpdateAttributeGroupResponse'

-> UpdateAttributeGroupResponse

Create a value of UpdateAttributeGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeGroup:UpdateAttributeGroup', updateAttributeGroupResponse_attributeGroup - The updated information of the attribute group.

$sel:httpStatus:UpdateAttributeGroupResponse', updateAttributeGroupResponse_httpStatus - The response's http status code.

Response Lenses

updateAttributeGroupResponse_attributeGroup :: Lens' UpdateAttributeGroupResponse (Maybe AttributeGroup)

The updated information of the attribute group.

updateAttributeGroupResponse_httpStatus :: Lens' UpdateAttributeGroupResponse Int

The response's http status code.