Network.AWS.ServiceQuotas.Lens

Description

Operations

ListServices

listServices_nextToken :: Lens' ListServices (Maybe Text)

The token for the next page of results.

listServices_maxResults :: Lens' ListServices (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

listServicesResponse_nextToken :: Lens' ListServicesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listServicesResponse_services :: Lens' ListServicesResponse (Maybe [ServiceInfo])

Information about the services.

listServicesResponse_httpStatus :: Lens' ListServicesResponse Int

The response's http status code.

ListAWSDefaultServiceQuotas

listAWSDefaultServiceQuotas_nextToken :: Lens' ListAWSDefaultServiceQuotas (Maybe Text)

The token for the next page of results.

listAWSDefaultServiceQuotas_maxResults :: Lens' ListAWSDefaultServiceQuotas (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

listAWSDefaultServiceQuotas_serviceCode :: Lens' ListAWSDefaultServiceQuotas Text

The service identifier.

listAWSDefaultServiceQuotasResponse_nextToken :: Lens' ListAWSDefaultServiceQuotasResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listAWSDefaultServiceQuotasResponse_quotas :: Lens' ListAWSDefaultServiceQuotasResponse (Maybe [ServiceQuota])

Information about the quotas.

listAWSDefaultServiceQuotasResponse_httpStatus :: Lens' ListAWSDefaultServiceQuotasResponse Int

The response's http status code.

GetAssociationForServiceQuotaTemplate

getAssociationForServiceQuotaTemplateResponse_serviceQuotaTemplateAssociationStatus :: Lens' GetAssociationForServiceQuotaTemplateResponse (Maybe ServiceQuotaTemplateAssociationStatus)

The association status. If the status is ASSOCIATED, the quota increase requests in the template are automatically applied to new accounts in your organization.

getAssociationForServiceQuotaTemplateResponse_httpStatus :: Lens' GetAssociationForServiceQuotaTemplateResponse Int

The response's http status code.

AssociateServiceQuotaTemplate

associateServiceQuotaTemplateResponse_httpStatus :: Lens' AssociateServiceQuotaTemplateResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceARN :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) for the applied quota for which you want to list tags. You can get this information by using the Service Quotas console, or by listing the quotas using the list-service-quotas AWS CLI command or the ListServiceQuotas AWS API operation.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

A complex data type that contains zero or more tag elements.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

GetServiceQuota

getServiceQuota_serviceCode :: Lens' GetServiceQuota Text

The service identifier.

getServiceQuota_quotaCode :: Lens' GetServiceQuota Text

The quota identifier.

getServiceQuotaResponse_quota :: Lens' GetServiceQuotaResponse (Maybe ServiceQuota)

Information about the quota.

getServiceQuotaResponse_httpStatus :: Lens' GetServiceQuotaResponse Int

The response's http status code.

PutServiceQuotaIncreaseRequestIntoTemplate

putServiceQuotaIncreaseRequestIntoTemplate_quotaCode :: Lens' PutServiceQuotaIncreaseRequestIntoTemplate Text

The quota identifier.

putServiceQuotaIncreaseRequestIntoTemplate_serviceCode :: Lens' PutServiceQuotaIncreaseRequestIntoTemplate Text

The service identifier.

putServiceQuotaIncreaseRequestIntoTemplate_awsRegion :: Lens' PutServiceQuotaIncreaseRequestIntoTemplate Text

The AWS Region.

putServiceQuotaIncreaseRequestIntoTemplate_desiredValue :: Lens' PutServiceQuotaIncreaseRequestIntoTemplate Double

The new, increased value for the quota.

putServiceQuotaIncreaseRequestIntoTemplateResponse_serviceQuotaIncreaseRequestInTemplate :: Lens' PutServiceQuotaIncreaseRequestIntoTemplateResponse (Maybe ServiceQuotaIncreaseRequestInTemplate)

Information about the quota increase request.

putServiceQuotaIncreaseRequestIntoTemplateResponse_httpStatus :: Lens' PutServiceQuotaIncreaseRequestIntoTemplateResponse Int

The response's http status code.

RequestServiceQuotaIncrease

requestServiceQuotaIncrease_serviceCode :: Lens' RequestServiceQuotaIncrease Text

The service identifier.

requestServiceQuotaIncrease_quotaCode :: Lens' RequestServiceQuotaIncrease Text

The quota identifier.

requestServiceQuotaIncrease_desiredValue :: Lens' RequestServiceQuotaIncrease Double

The new, increased value for the quota.

requestServiceQuotaIncreaseResponse_requestedQuota :: Lens' RequestServiceQuotaIncreaseResponse (Maybe RequestedServiceQuotaChange)

Information about the quota increase request.

requestServiceQuotaIncreaseResponse_httpStatus :: Lens' RequestServiceQuotaIncreaseResponse Int

The response's http status code.

GetServiceQuotaIncreaseRequestFromTemplate

getServiceQuotaIncreaseRequestFromTemplate_serviceCode :: Lens' GetServiceQuotaIncreaseRequestFromTemplate Text

The service identifier.

getServiceQuotaIncreaseRequestFromTemplate_quotaCode :: Lens' GetServiceQuotaIncreaseRequestFromTemplate Text

The quota identifier.

getServiceQuotaIncreaseRequestFromTemplate_awsRegion :: Lens' GetServiceQuotaIncreaseRequestFromTemplate Text

The AWS Region.

getServiceQuotaIncreaseRequestFromTemplateResponse_serviceQuotaIncreaseRequestInTemplate :: Lens' GetServiceQuotaIncreaseRequestFromTemplateResponse (Maybe ServiceQuotaIncreaseRequestInTemplate)

Information about the quota increase request.

getServiceQuotaIncreaseRequestFromTemplateResponse_httpStatus :: Lens' GetServiceQuotaIncreaseRequestFromTemplateResponse Int

The response's http status code.

DisassociateServiceQuotaTemplate

disassociateServiceQuotaTemplateResponse_httpStatus :: Lens' DisassociateServiceQuotaTemplateResponse Int

The response's http status code.

DeleteServiceQuotaIncreaseRequestFromTemplate

deleteServiceQuotaIncreaseRequestFromTemplate_serviceCode :: Lens' DeleteServiceQuotaIncreaseRequestFromTemplate Text

The service identifier.

deleteServiceQuotaIncreaseRequestFromTemplate_quotaCode :: Lens' DeleteServiceQuotaIncreaseRequestFromTemplate Text

The quota identifier.

deleteServiceQuotaIncreaseRequestFromTemplate_awsRegion :: Lens' DeleteServiceQuotaIncreaseRequestFromTemplate Text

The AWS Region.

deleteServiceQuotaIncreaseRequestFromTemplateResponse_httpStatus :: Lens' DeleteServiceQuotaIncreaseRequestFromTemplateResponse Int

The response's http status code.

ListServiceQuotas

listServiceQuotas_nextToken :: Lens' ListServiceQuotas (Maybe Text)

The token for the next page of results.

listServiceQuotas_maxResults :: Lens' ListServiceQuotas (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

listServiceQuotas_serviceCode :: Lens' ListServiceQuotas Text

The service identifier.

listServiceQuotasResponse_nextToken :: Lens' ListServiceQuotasResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listServiceQuotasResponse_quotas :: Lens' ListServiceQuotasResponse (Maybe [ServiceQuota])

Information about the quotas.

listServiceQuotasResponse_httpStatus :: Lens' ListServiceQuotasResponse Int

The response's http status code.

ListRequestedServiceQuotaChangeHistory

listRequestedServiceQuotaChangeHistory_status :: Lens' ListRequestedServiceQuotaChangeHistory (Maybe RequestStatus)

The status of the quota increase request.

listRequestedServiceQuotaChangeHistory_nextToken :: Lens' ListRequestedServiceQuotaChangeHistory (Maybe Text)

The token for the next page of results.

listRequestedServiceQuotaChangeHistory_serviceCode :: Lens' ListRequestedServiceQuotaChangeHistory (Maybe Text)

The service identifier.

listRequestedServiceQuotaChangeHistory_maxResults :: Lens' ListRequestedServiceQuotaChangeHistory (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

listRequestedServiceQuotaChangeHistoryResponse_nextToken :: Lens' ListRequestedServiceQuotaChangeHistoryResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listRequestedServiceQuotaChangeHistoryResponse_requestedQuotas :: Lens' ListRequestedServiceQuotaChangeHistoryResponse (Maybe [RequestedServiceQuotaChange])

Information about the quota increase requests.

listRequestedServiceQuotaChangeHistoryResponse_httpStatus :: Lens' ListRequestedServiceQuotaChangeHistoryResponse Int

The response's http status code.

ListServiceQuotaIncreaseRequestsInTemplate

listServiceQuotaIncreaseRequestsInTemplate_nextToken :: Lens' ListServiceQuotaIncreaseRequestsInTemplate (Maybe Text)

The token for the next page of results.

listServiceQuotaIncreaseRequestsInTemplate_awsRegion :: Lens' ListServiceQuotaIncreaseRequestsInTemplate (Maybe Text)

The AWS Region.

listServiceQuotaIncreaseRequestsInTemplate_serviceCode :: Lens' ListServiceQuotaIncreaseRequestsInTemplate (Maybe Text)

The service identifier.

listServiceQuotaIncreaseRequestsInTemplate_maxResults :: Lens' ListServiceQuotaIncreaseRequestsInTemplate (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

listServiceQuotaIncreaseRequestsInTemplateResponse_serviceQuotaIncreaseRequestInTemplateList :: Lens' ListServiceQuotaIncreaseRequestsInTemplateResponse (Maybe [ServiceQuotaIncreaseRequestInTemplate])

Information about the quota increase requests.

listServiceQuotaIncreaseRequestsInTemplateResponse_nextToken :: Lens' ListServiceQuotaIncreaseRequestsInTemplateResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listServiceQuotaIncreaseRequestsInTemplateResponse_httpStatus :: Lens' ListServiceQuotaIncreaseRequestsInTemplateResponse Int

The response's http status code.

TagResource

tagResource_resourceARN :: Lens' TagResource Text

The Amazon Resource Name (ARN) for the applied quota. You can get this information by using the Service Quotas console, or by listing the quotas using the list-service-quotas AWS CLI command or the ListServiceQuotas AWS API operation.

tagResource_tags :: Lens' TagResource (NonEmpty Tag)

The tags that you want to add to the resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

ListRequestedServiceQuotaChangeHistoryByQuota

listRequestedServiceQuotaChangeHistoryByQuota_status :: Lens' ListRequestedServiceQuotaChangeHistoryByQuota (Maybe RequestStatus)

The status value of the quota increase request.

listRequestedServiceQuotaChangeHistoryByQuota_nextToken :: Lens' ListRequestedServiceQuotaChangeHistoryByQuota (Maybe Text)

The token for the next page of results.

listRequestedServiceQuotaChangeHistoryByQuota_maxResults :: Lens' ListRequestedServiceQuotaChangeHistoryByQuota (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

listRequestedServiceQuotaChangeHistoryByQuota_serviceCode :: Lens' ListRequestedServiceQuotaChangeHistoryByQuota Text

The service identifier.

listRequestedServiceQuotaChangeHistoryByQuota_quotaCode :: Lens' ListRequestedServiceQuotaChangeHistoryByQuota Text

The quota identifier.

listRequestedServiceQuotaChangeHistoryByQuotaResponse_nextToken :: Lens' ListRequestedServiceQuotaChangeHistoryByQuotaResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listRequestedServiceQuotaChangeHistoryByQuotaResponse_requestedQuotas :: Lens' ListRequestedServiceQuotaChangeHistoryByQuotaResponse (Maybe [RequestedServiceQuotaChange])

Information about the quota increase requests.

listRequestedServiceQuotaChangeHistoryByQuotaResponse_httpStatus :: Lens' ListRequestedServiceQuotaChangeHistoryByQuotaResponse Int

The response's http status code.

GetRequestedServiceQuotaChange

getRequestedServiceQuotaChange_requestId :: Lens' GetRequestedServiceQuotaChange Text

The ID of the quota increase request.

getRequestedServiceQuotaChangeResponse_requestedQuota :: Lens' GetRequestedServiceQuotaChangeResponse (Maybe RequestedServiceQuotaChange)

Information about the quota increase request.

getRequestedServiceQuotaChangeResponse_httpStatus :: Lens' GetRequestedServiceQuotaChangeResponse Int

The response's http status code.

UntagResource

untagResource_resourceARN :: Lens' UntagResource Text

The Amazon Resource Name (ARN) for the applied quota that you want to untag. You can get this information by using the Service Quotas console, or by listing the quotas using the list-service-quotas AWS CLI command or the ListServiceQuotas AWS API operation.

untagResource_tagKeys :: Lens' UntagResource [Text]

The keys of the tags that you want to remove from the resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

GetAWSDefaultServiceQuota

getAWSDefaultServiceQuota_serviceCode :: Lens' GetAWSDefaultServiceQuota Text

The service identifier.

getAWSDefaultServiceQuota_quotaCode :: Lens' GetAWSDefaultServiceQuota Text

The quota identifier.

getAWSDefaultServiceQuotaResponse_quota :: Lens' GetAWSDefaultServiceQuotaResponse (Maybe ServiceQuota)

Information about the quota.

getAWSDefaultServiceQuotaResponse_httpStatus :: Lens' GetAWSDefaultServiceQuotaResponse Int

The response's http status code.

Types

ErrorReason

errorReason_errorCode :: Lens' ErrorReason (Maybe ErrorCode)

Service Quotas returns the following error values:

DEPENDENCY_ACCESS_DENIED_ERROR - The caller does not have the required permissions to complete the action. To resolve the error, you must have permission to access the service or quota.

DEPENDENCY_THROTTLING_ERROR - The service is throttling Service Quotas.

DEPENDENCY_SERVICE_ERROR - The service is not available.

SERVICE_QUOTA_NOT_AVAILABLE_ERROR - There was an error in Service Quotas.

errorReason_errorMessage :: Lens' ErrorReason (Maybe Text)

The error message.

MetricInfo

metricInfo_metricDimensions :: Lens' MetricInfo (Maybe (HashMap Text Text))

The metric dimension. This is a name/value pair that is part of the identity of a metric.

metricInfo_metricName :: Lens' MetricInfo (Maybe Text)

The name of the metric.

metricInfo_metricStatisticRecommendation :: Lens' MetricInfo (Maybe Text)

The metric statistic that we recommend you use when determining quota usage.

metricInfo_metricNamespace :: Lens' MetricInfo (Maybe Text)

The namespace of the metric.

QuotaPeriod

quotaPeriod_periodUnit :: Lens' QuotaPeriod (Maybe PeriodUnit)

The time unit.

quotaPeriod_periodValue :: Lens' QuotaPeriod (Maybe Int)

The value.

RequestedServiceQuotaChange

requestedServiceQuotaChange_status :: Lens' RequestedServiceQuotaChange (Maybe RequestStatus)

The state of the quota increase request.

requestedServiceQuotaChange_lastUpdated :: Lens' RequestedServiceQuotaChange (Maybe UTCTime)

The date and time of the most recent change.

requestedServiceQuotaChange_globalQuota :: Lens' RequestedServiceQuotaChange (Maybe Bool)

Indicates whether the quota is global.

requestedServiceQuotaChange_created :: Lens' RequestedServiceQuotaChange (Maybe UTCTime)

The date and time when the quota increase request was received and the case ID was created.

requestedServiceQuotaChange_desiredValue :: Lens' RequestedServiceQuotaChange (Maybe Double)

The new, increased value for the quota.

requestedServiceQuotaChange_quotaArn :: Lens' RequestedServiceQuotaChange (Maybe Text)

The Amazon Resource Name (ARN) of the quota.

requestedServiceQuotaChange_caseId :: Lens' RequestedServiceQuotaChange (Maybe Text)

The case ID.

requestedServiceQuotaChange_serviceName :: Lens' RequestedServiceQuotaChange (Maybe Text)

The service name.

requestedServiceQuotaChange_id :: Lens' RequestedServiceQuotaChange (Maybe Text)

The unique identifier.

requestedServiceQuotaChange_serviceCode :: Lens' RequestedServiceQuotaChange (Maybe Text)

The service identifier.

requestedServiceQuotaChange_quotaCode :: Lens' RequestedServiceQuotaChange (Maybe Text)

The quota identifier.

requestedServiceQuotaChange_unit :: Lens' RequestedServiceQuotaChange (Maybe Text)

The unit of measurement.

requestedServiceQuotaChange_requester :: Lens' RequestedServiceQuotaChange (Maybe Text)

The IAM identity of the requester.

requestedServiceQuotaChange_quotaName :: Lens' RequestedServiceQuotaChange (Maybe Text)

The quota name.

ServiceInfo

serviceInfo_serviceName :: Lens' ServiceInfo (Maybe Text)

The service name.

serviceInfo_serviceCode :: Lens' ServiceInfo (Maybe Text)

The service identifier.

ServiceQuota

serviceQuota_globalQuota :: Lens' ServiceQuota (Maybe Bool)

Indicates whether the quota is global.

serviceQuota_period :: Lens' ServiceQuota (Maybe QuotaPeriod)

The period of time.

serviceQuota_value :: Lens' ServiceQuota (Maybe Double)

The quota value.

serviceQuota_quotaArn :: Lens' ServiceQuota (Maybe Text)

The Amazon Resource Name (ARN) of the quota.

serviceQuota_usageMetric :: Lens' ServiceQuota (Maybe MetricInfo)

Information about the measurement.

serviceQuota_errorReason :: Lens' ServiceQuota (Maybe ErrorReason)

The error code and error reason.

serviceQuota_adjustable :: Lens' ServiceQuota (Maybe Bool)

Indicates whether the quota value can be increased.

serviceQuota_serviceName :: Lens' ServiceQuota (Maybe Text)

The service name.

serviceQuota_serviceCode :: Lens' ServiceQuota (Maybe Text)

The service identifier.

serviceQuota_quotaCode :: Lens' ServiceQuota (Maybe Text)

The quota identifier.

serviceQuota_unit :: Lens' ServiceQuota (Maybe Text)

The unit of measurement.

serviceQuota_quotaName :: Lens' ServiceQuota (Maybe Text)

The quota name.

ServiceQuotaIncreaseRequestInTemplate

serviceQuotaIncreaseRequestInTemplate_globalQuota :: Lens' ServiceQuotaIncreaseRequestInTemplate (Maybe Bool)

Indicates whether the quota is global.

serviceQuotaIncreaseRequestInTemplate_desiredValue :: Lens' ServiceQuotaIncreaseRequestInTemplate (Maybe Double)

The new, increased value of the quota.

serviceQuotaIncreaseRequestInTemplate_serviceName :: Lens' ServiceQuotaIncreaseRequestInTemplate (Maybe Text)

The service name.

serviceQuotaIncreaseRequestInTemplate_awsRegion :: Lens' ServiceQuotaIncreaseRequestInTemplate (Maybe Text)

The AWS Region.

serviceQuotaIncreaseRequestInTemplate_serviceCode :: Lens' ServiceQuotaIncreaseRequestInTemplate (Maybe Text)

The service identifier.

serviceQuotaIncreaseRequestInTemplate_quotaCode :: Lens' ServiceQuotaIncreaseRequestInTemplate (Maybe Text)

The quota identifier.

serviceQuotaIncreaseRequestInTemplate_unit :: Lens' ServiceQuotaIncreaseRequestInTemplate (Maybe Text)

The unit of measurement.

serviceQuotaIncreaseRequestInTemplate_quotaName :: Lens' ServiceQuotaIncreaseRequestInTemplate (Maybe Text)

The quota name.

Tag

tag_key :: Lens' Tag Text

A string that contains a tag key. The string length should be between 1 and 128 characters. Valid characters include a-z, A-Z, 0-9, space, and the special characters _ - . : / = + @.

tag_value :: Lens' Tag Text

A string that contains an optional tag value. The string length should be between 0 and 256 characters. Valid characters include a-z, A-Z, 0-9, space, and the special characters _ - . : / = + @.