Network.AWS.ServiceQuotas.Types.RequestedServiceQuotaChange

Description

Documentation

data RequestedServiceQuotaChange

Information about a quota increase request.

See: newRequestedServiceQuotaChange smart constructor.

Constructors

RequestedServiceQuotaChange'

Fields

status :: Maybe RequestStatus

The state of the quota increase request.

lastUpdated :: Maybe POSIX

The date and time of the most recent change.

globalQuota :: Maybe Bool

Indicates whether the quota is global.

created :: Maybe POSIX

The date and time when the quota increase request was received and the case ID was created.

desiredValue :: Maybe Double

The new, increased value for the quota.

quotaArn :: Maybe Text

The Amazon Resource Name (ARN) of the quota.

caseId :: Maybe Text

The case ID.

serviceName :: Maybe Text

The service name.

id :: Maybe Text

The unique identifier.

serviceCode :: Maybe Text

The service identifier.

quotaCode :: Maybe Text

The quota identifier.

unit :: Maybe Text

The unit of measurement.

requester :: Maybe Text

The IAM identity of the requester.

quotaName :: Maybe Text

The quota name.

newRequestedServiceQuotaChange :: RequestedServiceQuotaChange

Create a value of RequestedServiceQuotaChange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:RequestedServiceQuotaChange', requestedServiceQuotaChange_status - The state of the quota increase request.

$sel:lastUpdated:RequestedServiceQuotaChange', requestedServiceQuotaChange_lastUpdated - The date and time of the most recent change.

$sel:globalQuota:RequestedServiceQuotaChange', requestedServiceQuotaChange_globalQuota - Indicates whether the quota is global.

$sel:created:RequestedServiceQuotaChange', requestedServiceQuotaChange_created - The date and time when the quota increase request was received and the case ID was created.

$sel:desiredValue:RequestedServiceQuotaChange', requestedServiceQuotaChange_desiredValue - The new, increased value for the quota.

$sel:quotaArn:RequestedServiceQuotaChange', requestedServiceQuotaChange_quotaArn - The Amazon Resource Name (ARN) of the quota.

$sel:caseId:RequestedServiceQuotaChange', requestedServiceQuotaChange_caseId - The case ID.

$sel:serviceName:RequestedServiceQuotaChange', requestedServiceQuotaChange_serviceName - The service name.

$sel:id:RequestedServiceQuotaChange', requestedServiceQuotaChange_id - The unique identifier.

$sel:serviceCode:RequestedServiceQuotaChange', requestedServiceQuotaChange_serviceCode - The service identifier.

$sel:quotaCode:RequestedServiceQuotaChange', requestedServiceQuotaChange_quotaCode - The quota identifier.

$sel:unit:RequestedServiceQuotaChange', requestedServiceQuotaChange_unit - The unit of measurement.

$sel:requester:RequestedServiceQuotaChange', requestedServiceQuotaChange_requester - The IAM identity of the requester.

$sel:quotaName:RequestedServiceQuotaChange', requestedServiceQuotaChange_quotaName - The quota name.

requestedServiceQuotaChange_status :: Lens' RequestedServiceQuotaChange (Maybe RequestStatus)

The state of the quota increase request.

requestedServiceQuotaChange_lastUpdated :: Lens' RequestedServiceQuotaChange (Maybe UTCTime)

The date and time of the most recent change.

requestedServiceQuotaChange_globalQuota :: Lens' RequestedServiceQuotaChange (Maybe Bool)

Indicates whether the quota is global.

requestedServiceQuotaChange_created :: Lens' RequestedServiceQuotaChange (Maybe UTCTime)

The date and time when the quota increase request was received and the case ID was created.

requestedServiceQuotaChange_desiredValue :: Lens' RequestedServiceQuotaChange (Maybe Double)

The new, increased value for the quota.

requestedServiceQuotaChange_quotaArn :: Lens' RequestedServiceQuotaChange (Maybe Text)

The Amazon Resource Name (ARN) of the quota.

requestedServiceQuotaChange_caseId :: Lens' RequestedServiceQuotaChange (Maybe Text)

The case ID.

requestedServiceQuotaChange_serviceName :: Lens' RequestedServiceQuotaChange (Maybe Text)

The service name.

requestedServiceQuotaChange_id :: Lens' RequestedServiceQuotaChange (Maybe Text)

The unique identifier.

requestedServiceQuotaChange_serviceCode :: Lens' RequestedServiceQuotaChange (Maybe Text)

The service identifier.

requestedServiceQuotaChange_quotaCode :: Lens' RequestedServiceQuotaChange (Maybe Text)

The quota identifier.

requestedServiceQuotaChange_unit :: Lens' RequestedServiceQuotaChange (Maybe Text)

The unit of measurement.

requestedServiceQuotaChange_requester :: Lens' RequestedServiceQuotaChange (Maybe Text)

The IAM identity of the requester.

requestedServiceQuotaChange_quotaName :: Lens' RequestedServiceQuotaChange (Maybe Text)

The quota name.