Network.AWS.Shield.AssociateHealthCheck

Description

Adds health-based detection to the Shield Advanced protection for a resource. Shield Advanced health-based detection uses the health of your Amazon Web Services resource to improve responsiveness and accuracy in attack detection and mitigation.

You define the health check in Route 53 and then associate it with your Shield Advanced protection. For more information, see Shield Advanced Health-Based Detection in the WAF Developer Guide.

Creating a Request

data AssociateHealthCheck

See: newAssociateHealthCheck smart constructor.

Constructors

AssociateHealthCheck'

Fields

protectionId :: Text

The unique identifier (ID) for the Protection object to add the health check association to.

healthCheckArn :: Text

The Amazon Resource Name (ARN) of the health check to associate with the protection.

newAssociateHealthCheck

Arguments

:: Text
$sel:protectionId:AssociateHealthCheck'

-> Text
$sel:healthCheckArn:AssociateHealthCheck'

-> AssociateHealthCheck

Create a value of AssociateHealthCheck with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protectionId:AssociateHealthCheck', associateHealthCheck_protectionId - The unique identifier (ID) for the Protection object to add the health check association to.

$sel:healthCheckArn:AssociateHealthCheck', associateHealthCheck_healthCheckArn - The Amazon Resource Name (ARN) of the health check to associate with the protection.

Request Lenses

associateHealthCheck_protectionId :: Lens' AssociateHealthCheck Text

The unique identifier (ID) for the Protection object to add the health check association to.

associateHealthCheck_healthCheckArn :: Lens' AssociateHealthCheck Text

The Amazon Resource Name (ARN) of the health check to associate with the protection.

Destructuring the Response

data AssociateHealthCheckResponse

See: newAssociateHealthCheckResponse smart constructor.

Constructors

AssociateHealthCheckResponse'

Fields

httpStatus :: Int

The response's http status code.

newAssociateHealthCheckResponse

Arguments

:: Int
$sel:httpStatus:AssociateHealthCheckResponse'

-> AssociateHealthCheckResponse

Create a value of AssociateHealthCheckResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AssociateHealthCheckResponse', associateHealthCheckResponse_httpStatus - The response's http status code.

Response Lenses

associateHealthCheckResponse_httpStatus :: Lens' AssociateHealthCheckResponse Int

The response's http status code.