Network.AWS.Shield.AssociateProactiveEngagementDetails

Description

Initializes proactive engagement and sets the list of contacts for the Shield Response Team (SRT) to use. You must provide at least one phone number in the emergency contact list.

After you have initialized proactive engagement using this call, to disable or enable proactive engagement, use the calls DisableProactiveEngagement and EnableProactiveEngagement.

This call defines the list of email addresses and phone numbers that the SRT can use to contact you for escalations to the SRT and to initiate proactive customer support.

The contacts that you provide in the request replace any contacts that were already defined. If you already have contacts defined and want to use them, retrieve the list using DescribeEmergencyContactSettings and then provide it to this call.

Creating a Request

data AssociateProactiveEngagementDetails

See: newAssociateProactiveEngagementDetails smart constructor.

Constructors

AssociateProactiveEngagementDetails'

Fields

emergencyContactList :: [EmergencyContact]

A list of email addresses and phone numbers that the Shield Response Team (SRT) can use to contact you for escalations to the SRT and to initiate proactive customer support.

To enable proactive engagement, the contact list must include at least one phone number.

The contacts that you provide here replace any contacts that were already defined. If you already have contacts defined and want to use them, retrieve the list using DescribeEmergencyContactSettings and then provide it here.

newAssociateProactiveEngagementDetails :: AssociateProactiveEngagementDetails

Create a value of AssociateProactiveEngagementDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:emergencyContactList:AssociateProactiveEngagementDetails', associateProactiveEngagementDetails_emergencyContactList - A list of email addresses and phone numbers that the Shield Response Team (SRT) can use to contact you for escalations to the SRT and to initiate proactive customer support.

To enable proactive engagement, the contact list must include at least one phone number.

The contacts that you provide here replace any contacts that were already defined. If you already have contacts defined and want to use them, retrieve the list using DescribeEmergencyContactSettings and then provide it here.

Request Lenses

associateProactiveEngagementDetails_emergencyContactList :: Lens' AssociateProactiveEngagementDetails [EmergencyContact]

A list of email addresses and phone numbers that the Shield Response Team (SRT) can use to contact you for escalations to the SRT and to initiate proactive customer support.

To enable proactive engagement, the contact list must include at least one phone number.

The contacts that you provide here replace any contacts that were already defined. If you already have contacts defined and want to use them, retrieve the list using DescribeEmergencyContactSettings and then provide it here.

Destructuring the Response

data AssociateProactiveEngagementDetailsResponse

See: newAssociateProactiveEngagementDetailsResponse smart constructor.

Constructors

AssociateProactiveEngagementDetailsResponse'

Fields

httpStatus :: Int

The response's http status code.

newAssociateProactiveEngagementDetailsResponse

Arguments

:: Int
$sel:httpStatus:AssociateProactiveEngagementDetailsResponse'

-> AssociateProactiveEngagementDetailsResponse

Create a value of AssociateProactiveEngagementDetailsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AssociateProactiveEngagementDetailsResponse', associateProactiveEngagementDetailsResponse_httpStatus - The response's http status code.

Response Lenses

associateProactiveEngagementDetailsResponse_httpStatus :: Lens' AssociateProactiveEngagementDetailsResponse Int

The response's http status code.