Network.AWS.Shield.DescribeEmergencyContactSettings

Description

A list of email addresses and phone numbers that the Shield Response Team (SRT) can use to contact you if you have proactive engagement enabled, for escalations to the SRT and to initiate proactive customer support.

Creating a Request

data DescribeEmergencyContactSettings

See: newDescribeEmergencyContactSettings smart constructor.

Constructors

DescribeEmergencyContactSettings'

newDescribeEmergencyContactSettings :: DescribeEmergencyContactSettings

Create a value of DescribeEmergencyContactSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data DescribeEmergencyContactSettingsResponse

See: newDescribeEmergencyContactSettingsResponse smart constructor.

Constructors

DescribeEmergencyContactSettingsResponse'

Fields

emergencyContactList :: Maybe [EmergencyContact]

A list of email addresses and phone numbers that the Shield Response Team (SRT) can use to contact you if you have proactive engagement enabled, for escalations to the SRT and to initiate proactive customer support.

httpStatus :: Int

The response's http status code.

newDescribeEmergencyContactSettingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeEmergencyContactSettingsResponse'

-> DescribeEmergencyContactSettingsResponse

Create a value of DescribeEmergencyContactSettingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:emergencyContactList:DescribeEmergencyContactSettingsResponse', describeEmergencyContactSettingsResponse_emergencyContactList - A list of email addresses and phone numbers that the Shield Response Team (SRT) can use to contact you if you have proactive engagement enabled, for escalations to the SRT and to initiate proactive customer support.

$sel:httpStatus:DescribeEmergencyContactSettingsResponse', describeEmergencyContactSettingsResponse_httpStatus - The response's http status code.

Response Lenses

describeEmergencyContactSettingsResponse_emergencyContactList :: Lens' DescribeEmergencyContactSettingsResponse (Maybe [EmergencyContact])

A list of email addresses and phone numbers that the Shield Response Team (SRT) can use to contact you if you have proactive engagement enabled, for escalations to the SRT and to initiate proactive customer support.

describeEmergencyContactSettingsResponse_httpStatus :: Lens' DescribeEmergencyContactSettingsResponse Int

The response's http status code.