Network.AWS.Shield.DescribeProtectionGroup

Description

Returns the specification for the specified protection group.

Creating a Request

data DescribeProtectionGroup

See: newDescribeProtectionGroup smart constructor.

Constructors

DescribeProtectionGroup'

Fields

protectionGroupId :: Text

The name of the protection group. You use this to identify the protection group in lists and to manage the protection group, for example to update, delete, or describe it.

newDescribeProtectionGroup

Arguments

:: Text
$sel:protectionGroupId:DescribeProtectionGroup'

-> DescribeProtectionGroup

Create a value of DescribeProtectionGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protectionGroupId:DescribeProtectionGroup', describeProtectionGroup_protectionGroupId - The name of the protection group. You use this to identify the protection group in lists and to manage the protection group, for example to update, delete, or describe it.

Request Lenses

describeProtectionGroup_protectionGroupId :: Lens' DescribeProtectionGroup Text

The name of the protection group. You use this to identify the protection group in lists and to manage the protection group, for example to update, delete, or describe it.

Destructuring the Response

data DescribeProtectionGroupResponse

See: newDescribeProtectionGroupResponse smart constructor.

Constructors

DescribeProtectionGroupResponse'

Fields

httpStatus :: Int

The response's http status code.

protectionGroup :: ProtectionGroup

A grouping of protected resources that you and Shield Advanced can monitor as a collective. This resource grouping improves the accuracy of detection and reduces false positives.

newDescribeProtectionGroupResponse

Arguments

:: Int
$sel:httpStatus:DescribeProtectionGroupResponse'

-> ProtectionGroup
$sel:protectionGroup:DescribeProtectionGroupResponse'

-> DescribeProtectionGroupResponse

Create a value of DescribeProtectionGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeProtectionGroupResponse', describeProtectionGroupResponse_httpStatus - The response's http status code.

$sel:protectionGroup:DescribeProtectionGroupResponse', describeProtectionGroupResponse_protectionGroup - A grouping of protected resources that you and Shield Advanced can monitor as a collective. This resource grouping improves the accuracy of detection and reduces false positives.

Response Lenses

describeProtectionGroupResponse_httpStatus :: Lens' DescribeProtectionGroupResponse Int

The response's http status code.

describeProtectionGroupResponse_protectionGroup :: Lens' DescribeProtectionGroupResponse ProtectionGroup

A grouping of protected resources that you and Shield Advanced can monitor as a collective. This resource grouping improves the accuracy of detection and reduces false positives.