Network.AWS.Shield.Lens

Description

Operations

AssociateDRTLogBucket

associateDRTLogBucket_logBucket :: Lens' AssociateDRTLogBucket Text

The Amazon S3 bucket that contains the logs that you want to share.

associateDRTLogBucketResponse_httpStatus :: Lens' AssociateDRTLogBucketResponse Int

The response's http status code.

DisassociateDRTRole

disassociateDRTRoleResponse_httpStatus :: Lens' DisassociateDRTRoleResponse Int

The response's http status code.

CreateSubscription

createSubscriptionResponse_httpStatus :: Lens' CreateSubscriptionResponse Int

The response's http status code.

ListProtections

listProtections_nextToken :: Lens' ListProtections (Maybe Text)

The ListProtectionsRequest.NextToken value from a previous call to ListProtections. Pass null if this is the first call.

listProtections_maxResults :: Lens' ListProtections (Maybe Natural)

The maximum number of Protection objects to return. If you leave this blank, Shield Advanced returns the first 20 results.

This is a maximum value. Shield Advanced might return the results in smaller batches. That is, the number of objects returned could be less than MaxResults, even if there are still more objects yet to return. If there are more objects to return, Shield Advanced returns a value in NextToken that you can use in your next request, to get the next batch of objects.

listProtectionsResponse_protections :: Lens' ListProtectionsResponse (Maybe [Protection])

The array of enabled Protection objects.

listProtectionsResponse_nextToken :: Lens' ListProtectionsResponse (Maybe Text)

If you specify a value for MaxResults and you have more Protections than the value of MaxResults, Shield Advanced returns a NextToken value in the response that allows you to list another group of Protections. For the second and subsequent ListProtections requests, specify the value of NextToken from the previous response to get information about another batch of Protections.

Shield Advanced might return the list of Protection objects in batches smaller than the number specified by MaxResults. If there are more Protection objects to return, Shield Advanced will always also return a NextToken.

listProtectionsResponse_httpStatus :: Lens' ListProtectionsResponse Int

The response's http status code.

AssociateDRTRole

associateDRTRole_roleArn :: Lens' AssociateDRTRole Text

The Amazon Resource Name (ARN) of the role the SRT will use to access your Amazon Web Services account.

Prior to making the AssociateDRTRole request, you must attach the AWSShieldDRTAccessPolicy managed policy to this role. For more information see Attaching and Detaching IAM Policies.

associateDRTRoleResponse_httpStatus :: Lens' AssociateDRTRoleResponse Int

The response's http status code.

UpdateSubscription

updateSubscription_autoRenew :: Lens' UpdateSubscription (Maybe AutoRenew)

When you initally create a subscription, AutoRenew is set to ENABLED. If ENABLED, the subscription will be automatically renewed at the end of the existing subscription period. You can change this by submitting an UpdateSubscription request. If the UpdateSubscription request does not included a value for AutoRenew, the existing value for AutoRenew remains unchanged.

updateSubscriptionResponse_httpStatus :: Lens' UpdateSubscriptionResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceARN :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the resource to get tags for.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

A list of tag key and value pairs associated with the specified resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

DisassociateDRTLogBucket

disassociateDRTLogBucket_logBucket :: Lens' DisassociateDRTLogBucket Text

The Amazon S3 bucket that contains the logs that you want to share.

disassociateDRTLogBucketResponse_httpStatus :: Lens' DisassociateDRTLogBucketResponse Int

The response's http status code.

AssociateProactiveEngagementDetails

associateProactiveEngagementDetails_emergencyContactList :: Lens' AssociateProactiveEngagementDetails [EmergencyContact]

A list of email addresses and phone numbers that the Shield Response Team (SRT) can use to contact you for escalations to the SRT and to initiate proactive customer support.

To enable proactive engagement, the contact list must include at least one phone number.

The contacts that you provide here replace any contacts that were already defined. If you already have contacts defined and want to use them, retrieve the list using DescribeEmergencyContactSettings and then provide it here.

associateProactiveEngagementDetailsResponse_httpStatus :: Lens' AssociateProactiveEngagementDetailsResponse Int

The response's http status code.

DescribeAttack

describeAttack_attackId :: Lens' DescribeAttack Text

The unique identifier (ID) for the attack that to be described.

describeAttackResponse_attack :: Lens' DescribeAttackResponse (Maybe AttackDetail)

The attack that is described.

describeAttackResponse_httpStatus :: Lens' DescribeAttackResponse Int

The response's http status code.

ListProtectionGroups

listProtectionGroups_nextToken :: Lens' ListProtectionGroups (Maybe Text)

The next token value from a previous call to ListProtectionGroups. Pass null if this is the first call.

listProtectionGroups_maxResults :: Lens' ListProtectionGroups (Maybe Natural)

The maximum number of ProtectionGroup objects to return. If you leave this blank, Shield Advanced returns the first 20 results.

This is a maximum value. Shield Advanced might return the results in smaller batches. That is, the number of objects returned could be less than MaxResults, even if there are still more objects yet to return. If there are more objects to return, Shield Advanced returns a value in NextToken that you can use in your next request, to get the next batch of objects.

listProtectionGroupsResponse_nextToken :: Lens' ListProtectionGroupsResponse (Maybe Text)

If you specify a value for MaxResults and you have more protection groups than the value of MaxResults, Shield Advanced returns this token that you can use in your next request, to get the next batch of objects.

listProtectionGroupsResponse_httpStatus :: Lens' ListProtectionGroupsResponse Int

The response's http status code.

listProtectionGroupsResponse_protectionGroups :: Lens' ListProtectionGroupsResponse [ProtectionGroup]

EnableProactiveEngagement

enableProactiveEngagementResponse_httpStatus :: Lens' EnableProactiveEngagementResponse Int

The response's http status code.

UpdateEmergencyContactSettings

updateEmergencyContactSettings_emergencyContactList :: Lens' UpdateEmergencyContactSettings (Maybe [EmergencyContact])

A list of email addresses and phone numbers that the Shield Response Team (SRT) can use to contact you if you have proactive engagement enabled, for escalations to the SRT and to initiate proactive customer support.

If you have proactive engagement enabled, the contact list must include at least one phone number.

updateEmergencyContactSettingsResponse_httpStatus :: Lens' UpdateEmergencyContactSettingsResponse Int

The response's http status code.

CreateProtectionGroup

createProtectionGroup_resourceType :: Lens' CreateProtectionGroup (Maybe ProtectedResourceType)

The resource type to include in the protection group. All protected resources of this type are included in the protection group. Newly protected resources of this type are automatically added to the group. You must set this when you set Pattern to BY_RESOURCE_TYPE and you must not set it for any other Pattern setting.

createProtectionGroup_members :: Lens' CreateProtectionGroup (Maybe [Text])

The Amazon Resource Names (ARNs) of the resources to include in the protection group. You must set this when you set Pattern to ARBITRARY and you must not set it for any other Pattern setting.

createProtectionGroup_tags :: Lens' CreateProtectionGroup (Maybe [Tag])

One or more tag key-value pairs for the protection group.

createProtectionGroup_protectionGroupId :: Lens' CreateProtectionGroup Text

The name of the protection group. You use this to identify the protection group in lists and to manage the protection group, for example to update, delete, or describe it.

createProtectionGroup_aggregation :: Lens' CreateProtectionGroup ProtectionGroupAggregation

Defines how Shield combines resource data for the group in order to detect, mitigate, and report events.

Sum - Use the total traffic across the group. This is a good choice for most cases. Examples include Elastic IP addresses for EC2 instances that scale manually or automatically.

Mean - Use the average of the traffic across the group. This is a good choice for resources that share traffic uniformly. Examples include accelerators and load balancers.

Max - Use the highest traffic from each resource. This is useful for resources that don't share traffic and for resources that share that traffic in a non-uniform way. Examples include Amazon CloudFront and origin resources for CloudFront distributions.

createProtectionGroup_pattern :: Lens' CreateProtectionGroup ProtectionGroupPattern

The criteria to use to choose the protected resources for inclusion in the group. You can include all resources that have protections, provide a list of resource Amazon Resource Names (ARNs), or include all resources of a specified resource type.

createProtectionGroupResponse_httpStatus :: Lens' CreateProtectionGroupResponse Int

The response's http status code.

DisableProactiveEngagement

disableProactiveEngagementResponse_httpStatus :: Lens' DisableProactiveEngagementResponse Int

The response's http status code.

DisassociateHealthCheck

disassociateHealthCheck_protectionId :: Lens' DisassociateHealthCheck Text

The unique identifier (ID) for the Protection object to remove the health check association from.

disassociateHealthCheck_healthCheckArn :: Lens' DisassociateHealthCheck Text

The Amazon Resource Name (ARN) of the health check that is associated with the protection.

disassociateHealthCheckResponse_httpStatus :: Lens' DisassociateHealthCheckResponse Int

The response's http status code.

ListResourcesInProtectionGroup

listResourcesInProtectionGroup_nextToken :: Lens' ListResourcesInProtectionGroup (Maybe Text)

The next token value from a previous call to ListResourcesInProtectionGroup. Pass null if this is the first call.

listResourcesInProtectionGroup_maxResults :: Lens' ListResourcesInProtectionGroup (Maybe Natural)

The maximum number of resource ARN objects to return. If you leave this blank, Shield Advanced returns the first 20 results.

This is a maximum value. Shield Advanced might return the results in smaller batches. That is, the number of objects returned could be less than MaxResults, even if there are still more objects yet to return. If there are more objects to return, Shield Advanced returns a value in NextToken that you can use in your next request, to get the next batch of objects.

listResourcesInProtectionGroup_protectionGroupId :: Lens' ListResourcesInProtectionGroup Text

The name of the protection group. You use this to identify the protection group in lists and to manage the protection group, for example to update, delete, or describe it.

listResourcesInProtectionGroupResponse_nextToken :: Lens' ListResourcesInProtectionGroupResponse (Maybe Text)

If you specify a value for MaxResults and you have more resources in the protection group than the value of MaxResults, Shield Advanced returns this token that you can use in your next request, to get the next batch of objects.

listResourcesInProtectionGroupResponse_httpStatus :: Lens' ListResourcesInProtectionGroupResponse Int

The response's http status code.

listResourcesInProtectionGroupResponse_resourceArns :: Lens' ListResourcesInProtectionGroupResponse [Text]

The Amazon Resource Names (ARNs) of the resources that are included in the protection group.

DescribeProtection

describeProtection_protectionId :: Lens' DescribeProtection (Maybe Text)

The unique identifier (ID) for the Protection object that is described. When submitting the DescribeProtection request you must provide either the ResourceArn or the ProtectionID, but not both.

describeProtection_resourceArn :: Lens' DescribeProtection (Maybe Text)

The ARN (Amazon Resource Name) of the Amazon Web Services resource for the Protection object that is described. When submitting the DescribeProtection request you must provide either the ResourceArn or the ProtectionID, but not both.

describeProtectionResponse_protection :: Lens' DescribeProtectionResponse (Maybe Protection)

The Protection object that is described.

describeProtectionResponse_httpStatus :: Lens' DescribeProtectionResponse Int

The response's http status code.

ListAttacks

listAttacks_startTime :: Lens' ListAttacks (Maybe TimeRange)

The start of the time period for the attacks. This is a timestamp type. The sample request above indicates a number type because the default used by WAF is Unix time in seconds. However any valid timestamp format is allowed.

listAttacks_resourceArns :: Lens' ListAttacks (Maybe [Text])

The ARN (Amazon Resource Name) of the resource that was attacked. If this is left blank, all applicable resources for this account will be included.

listAttacks_nextToken :: Lens' ListAttacks (Maybe Text)

The ListAttacksRequest.NextMarker value from a previous call to ListAttacksRequest. Pass null if this is the first call.

listAttacks_endTime :: Lens' ListAttacks (Maybe TimeRange)

The end of the time period for the attacks. This is a timestamp type. The sample request above indicates a number type because the default used by WAF is Unix time in seconds. However any valid timestamp format is allowed.

listAttacks_maxResults :: Lens' ListAttacks (Maybe Natural)

The maximum number of AttackSummary objects to return. If you leave this blank, Shield Advanced returns the first 20 results.

This is a maximum value. Shield Advanced might return the results in smaller batches. That is, the number of objects returned could be less than MaxResults, even if there are still more objects yet to return. If there are more objects to return, Shield Advanced returns a value in NextToken that you can use in your next request, to get the next batch of objects.

listAttacksResponse_attackSummaries :: Lens' ListAttacksResponse (Maybe [AttackSummary])

The attack information for the specified time range.

listAttacksResponse_nextToken :: Lens' ListAttacksResponse (Maybe Text)

The token returned by a previous call to indicate that there is more data available. If not null, more results are available. Pass this value for the NextMarker parameter in a subsequent call to ListAttacks to retrieve the next set of items.

Shield Advanced might return the list of AttackSummary objects in batches smaller than the number specified by MaxResults. If there are more attack summary objects to return, Shield Advanced will always also return a NextToken.

listAttacksResponse_httpStatus :: Lens' ListAttacksResponse Int

The response's http status code.

DescribeEmergencyContactSettings

describeEmergencyContactSettingsResponse_emergencyContactList :: Lens' DescribeEmergencyContactSettingsResponse (Maybe [EmergencyContact])

A list of email addresses and phone numbers that the Shield Response Team (SRT) can use to contact you if you have proactive engagement enabled, for escalations to the SRT and to initiate proactive customer support.

describeEmergencyContactSettingsResponse_httpStatus :: Lens' DescribeEmergencyContactSettingsResponse Int

The response's http status code.

CreateProtection

createProtection_tags :: Lens' CreateProtection (Maybe [Tag])

One or more tag key-value pairs for the Protection object that is created.

createProtection_name :: Lens' CreateProtection Text

Friendly name for the Protection you are creating.

createProtection_resourceArn :: Lens' CreateProtection Text

The ARN (Amazon Resource Name) of the resource to be protected.

The ARN should be in one of the following formats:

For an Application Load Balancer: arn:aws:elasticloadbalancing:region:account-id:loadbalancer/app/load-balancer-name/load-balancer-id

For an Elastic Load Balancer (Classic Load Balancer): arn:aws:elasticloadbalancing:region:account-id:loadbalancer/load-balancer-name

For an Amazon CloudFront distribution: arn:aws:cloudfront::account-id:distribution/distribution-id

For an Global Accelerator accelerator: arn:aws:globalaccelerator::account-id:accelerator/accelerator-id

For Amazon Route 53: arn:aws:route53:::hostedzone/hosted-zone-id

For an Elastic IP address: arn:aws:ec2:region:account-id:eip-allocation/allocation-id

createProtectionResponse_protectionId :: Lens' CreateProtectionResponse (Maybe Text)

The unique identifier (ID) for the Protection object that is created.

createProtectionResponse_httpStatus :: Lens' CreateProtectionResponse Int

The response's http status code.

DeleteProtection

deleteProtection_protectionId :: Lens' DeleteProtection Text

The unique identifier (ID) for the Protection object to be deleted.

deleteProtectionResponse_httpStatus :: Lens' DeleteProtectionResponse Int

The response's http status code.

GetSubscriptionState

getSubscriptionStateResponse_httpStatus :: Lens' GetSubscriptionStateResponse Int

The response's http status code.

getSubscriptionStateResponse_subscriptionState :: Lens' GetSubscriptionStateResponse SubscriptionState

The status of the subscription.

DeleteProtectionGroup

deleteProtectionGroup_protectionGroupId :: Lens' DeleteProtectionGroup Text

The name of the protection group. You use this to identify the protection group in lists and to manage the protection group, for example to update, delete, or describe it.

deleteProtectionGroupResponse_httpStatus :: Lens' DeleteProtectionGroupResponse Int

The response's http status code.

UpdateProtectionGroup

updateProtectionGroup_resourceType :: Lens' UpdateProtectionGroup (Maybe ProtectedResourceType)

The resource type to include in the protection group. All protected resources of this type are included in the protection group. You must set this when you set Pattern to BY_RESOURCE_TYPE and you must not set it for any other Pattern setting.

updateProtectionGroup_members :: Lens' UpdateProtectionGroup (Maybe [Text])

The Amazon Resource Names (ARNs) of the resources to include in the protection group. You must set this when you set Pattern to ARBITRARY and you must not set it for any other Pattern setting.

updateProtectionGroup_protectionGroupId :: Lens' UpdateProtectionGroup Text

The name of the protection group. You use this to identify the protection group in lists and to manage the protection group, for example to update, delete, or describe it.

updateProtectionGroup_aggregation :: Lens' UpdateProtectionGroup ProtectionGroupAggregation

Defines how Shield combines resource data for the group in order to detect, mitigate, and report events.

Sum - Use the total traffic across the group. This is a good choice for most cases. Examples include Elastic IP addresses for EC2 instances that scale manually or automatically.

Mean - Use the average of the traffic across the group. This is a good choice for resources that share traffic uniformly. Examples include accelerators and load balancers.

Max - Use the highest traffic from each resource. This is useful for resources that don't share traffic and for resources that share that traffic in a non-uniform way. Examples include Amazon CloudFront distributions and origin resources for CloudFront distributions.

updateProtectionGroup_pattern :: Lens' UpdateProtectionGroup ProtectionGroupPattern

The criteria to use to choose the protected resources for inclusion in the group. You can include all resources that have protections, provide a list of resource Amazon Resource Names (ARNs), or include all resources of a specified resource type.

updateProtectionGroupResponse_httpStatus :: Lens' UpdateProtectionGroupResponse Int

The response's http status code.

DescribeAttackStatistics

describeAttackStatisticsResponse_httpStatus :: Lens' DescribeAttackStatisticsResponse Int

The response's http status code.

describeAttackStatisticsResponse_timeRange :: Lens' DescribeAttackStatisticsResponse TimeRange

Undocumented member.

describeAttackStatisticsResponse_dataItems :: Lens' DescribeAttackStatisticsResponse [AttackStatisticsDataItem]

The data that describes the attacks detected during the time period.

TagResource

tagResource_resourceARN :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the resource that you want to add or update tags for.

tagResource_tags :: Lens' TagResource [Tag]

The tags that you want to modify or add to the resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

DescribeDRTAccess

describeDRTAccessResponse_logBucketList :: Lens' DescribeDRTAccessResponse (Maybe [Text])

The list of Amazon S3 buckets accessed by the SRT.

describeDRTAccessResponse_roleArn :: Lens' DescribeDRTAccessResponse (Maybe Text)

The Amazon Resource Name (ARN) of the role the SRT used to access your Amazon Web Services account.

describeDRTAccessResponse_httpStatus :: Lens' DescribeDRTAccessResponse Int

The response's http status code.

UntagResource

untagResource_resourceARN :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the resource that you want to remove tags from.

untagResource_tagKeys :: Lens' UntagResource [Text]

The tag key for each tag that you want to remove from the resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

DescribeSubscription

describeSubscriptionResponse_subscription :: Lens' DescribeSubscriptionResponse (Maybe Subscription)

The Shield Advanced subscription details for an account.

describeSubscriptionResponse_httpStatus :: Lens' DescribeSubscriptionResponse Int

The response's http status code.

AssociateHealthCheck

associateHealthCheck_protectionId :: Lens' AssociateHealthCheck Text

The unique identifier (ID) for the Protection object to add the health check association to.

associateHealthCheck_healthCheckArn :: Lens' AssociateHealthCheck Text

The Amazon Resource Name (ARN) of the health check to associate with the protection.

associateHealthCheckResponse_httpStatus :: Lens' AssociateHealthCheckResponse Int

The response's http status code.

DescribeProtectionGroup

describeProtectionGroup_protectionGroupId :: Lens' DescribeProtectionGroup Text

The name of the protection group. You use this to identify the protection group in lists and to manage the protection group, for example to update, delete, or describe it.

describeProtectionGroupResponse_httpStatus :: Lens' DescribeProtectionGroupResponse Int

The response's http status code.

describeProtectionGroupResponse_protectionGroup :: Lens' DescribeProtectionGroupResponse ProtectionGroup

A grouping of protected resources that you and Shield Advanced can monitor as a collective. This resource grouping improves the accuracy of detection and reduces false positives.

Types

AttackDetail

attackDetail_attackId :: Lens' AttackDetail (Maybe Text)

The unique identifier (ID) of the attack.

attackDetail_startTime :: Lens' AttackDetail (Maybe UTCTime)

The time the attack started, in Unix time in seconds. For more information see timestamp.

attackDetail_subResources :: Lens' AttackDetail (Maybe [SubResourceSummary])

If applicable, additional detail about the resource being attacked, for example, IP address or URL.

attackDetail_mitigations :: Lens' AttackDetail (Maybe [Mitigation])

List of mitigation actions taken for the attack.

attackDetail_attackProperties :: Lens' AttackDetail (Maybe [AttackProperty])

The array of objects that provide details of the Shield event.

For infrastructure layer events (L3 and L4 events) after January 25, 2021, you can view metrics for top contributors in Amazon CloudWatch metrics. For more information, see Shield metrics and alarms in the WAF Developer Guide.

attackDetail_attackCounters :: Lens' AttackDetail (Maybe [SummarizedCounter])

List of counters that describe the attack for the specified time period.

attackDetail_resourceArn :: Lens' AttackDetail (Maybe Text)

The ARN (Amazon Resource Name) of the resource that was attacked.

attackDetail_endTime :: Lens' AttackDetail (Maybe UTCTime)

The time the attack ended, in Unix time in seconds. For more information see timestamp.

AttackProperty

attackProperty_attackLayer :: Lens' AttackProperty (Maybe AttackLayer)

The type of Shield event that was observed. NETWORK indicates layer 3 and layer 4 events and APPLICATION indicates layer 7 events.

For infrastructure layer events (L3 and L4 events) after January 25, 2021, you can view metrics for top contributors in Amazon CloudWatch metrics. For more information, see Shield metrics and alarms in the WAF Developer Guide.

attackProperty_topContributors :: Lens' AttackProperty (Maybe [Contributor])

Contributor objects for the top five contributors to a Shield event.

attackProperty_attackPropertyIdentifier :: Lens' AttackProperty (Maybe AttackPropertyIdentifier)

Defines the Shield event property information that is provided. The WORDPRESS_PINGBACK_REFLECTOR and WORDPRESS_PINGBACK_SOURCE values are valid only for WordPress reflective pingback events.

attackProperty_total :: Lens' AttackProperty (Maybe Integer)

The total contributions made to this Shield event by all contributors.

attackProperty_unit :: Lens' AttackProperty (Maybe Unit)

The unit used for the Contributor Value property.

AttackStatisticsDataItem

attackStatisticsDataItem_attackVolume :: Lens' AttackStatisticsDataItem (Maybe AttackVolume)

Information about the volume of attacks during the time period. If the accompanying AttackCount is zero, this setting might be empty.

attackStatisticsDataItem_attackCount :: Lens' AttackStatisticsDataItem Integer

The number of attacks detected during the time period. This is always present, but might be zero.

AttackSummary

attackSummary_attackVectors :: Lens' AttackSummary (Maybe [AttackVectorDescription])

The list of attacks for a specified time period.

attackSummary_attackId :: Lens' AttackSummary (Maybe Text)

The unique identifier (ID) of the attack.

attackSummary_startTime :: Lens' AttackSummary (Maybe UTCTime)

The start time of the attack, in Unix time in seconds. For more information see timestamp.

attackSummary_resourceArn :: Lens' AttackSummary (Maybe Text)

The ARN (Amazon Resource Name) of the resource that was attacked.

attackSummary_endTime :: Lens' AttackSummary (Maybe UTCTime)

The end time of the attack, in Unix time in seconds. For more information see timestamp.

AttackVectorDescription

attackVectorDescription_vectorType :: Lens' AttackVectorDescription Text

The attack type. Valid values:

UDP_TRAFFIC

UDP_FRAGMENT

GENERIC_UDP_REFLECTION

DNS_REFLECTION

NTP_REFLECTION

CHARGEN_REFLECTION

SSDP_REFLECTION

PORT_MAPPER

RIP_REFLECTION

SNMP_REFLECTION

MSSQL_REFLECTION

NET_BIOS_REFLECTION

SYN_FLOOD

ACK_FLOOD

REQUEST_FLOOD

HTTP_REFLECTION

UDS_REFLECTION

MEMCACHED_REFLECTION

AttackVolume

attackVolume_packetsPerSecond :: Lens' AttackVolume (Maybe AttackVolumeStatistics)

A statistics object that uses packets per second as the unit. This is included for network level attacks.

attackVolume_requestsPerSecond :: Lens' AttackVolume (Maybe AttackVolumeStatistics)

A statistics object that uses requests per second as the unit. This is included for application level attacks, and is only available for accounts that are subscribed to Shield Advanced.

attackVolume_bitsPerSecond :: Lens' AttackVolume (Maybe AttackVolumeStatistics)

A statistics object that uses bits per second as the unit. This is included for network level attacks.

AttackVolumeStatistics

attackVolumeStatistics_max :: Lens' AttackVolumeStatistics Double

The maximum attack volume observed for the given unit.

Contributor

contributor_value :: Lens' Contributor (Maybe Integer)

The contribution of this contributor expressed in Protection units. For example 10,000.

contributor_name :: Lens' Contributor (Maybe Text)

The name of the contributor. This is dependent on the AttackPropertyIdentifier. For example, if the AttackPropertyIdentifier is SOURCE_COUNTRY, the Name could be United States.

EmergencyContact

emergencyContact_phoneNumber :: Lens' EmergencyContact (Maybe Text)

The phone number for the contact.

emergencyContact_contactNotes :: Lens' EmergencyContact (Maybe Text)

Additional notes regarding the contact.

emergencyContact_emailAddress :: Lens' EmergencyContact Text

The email address for the contact.

Limit

limit_max :: Lens' Limit (Maybe Integer)

The maximum number of protections that can be created for the specified Type.

limit_type :: Lens' Limit (Maybe Text)

The type of protection.

Mitigation

mitigation_mitigationName :: Lens' Mitigation (Maybe Text)

The name of the mitigation taken for this attack.

Protection

protection_protectionArn :: Lens' Protection (Maybe Text)

The ARN (Amazon Resource Name) of the protection.

protection_healthCheckIds :: Lens' Protection (Maybe [Text])

The unique identifier (ID) for the Route 53 health check that's associated with the protection.

protection_resourceArn :: Lens' Protection (Maybe Text)

The ARN (Amazon Resource Name) of the Amazon Web Services resource that is protected.

protection_name :: Lens' Protection (Maybe Text)

The name of the protection. For example, My CloudFront distributions.

protection_id :: Lens' Protection (Maybe Text)

The unique identifier (ID) of the protection.

ProtectionGroup

protectionGroup_resourceType :: Lens' ProtectionGroup (Maybe ProtectedResourceType)

The resource type to include in the protection group. All protected resources of this type are included in the protection group. You must set this when you set Pattern to BY_RESOURCE_TYPE and you must not set it for any other Pattern setting.

protectionGroup_protectionGroupArn :: Lens' ProtectionGroup (Maybe Text)

The ARN (Amazon Resource Name) of the protection group.

protectionGroup_protectionGroupId :: Lens' ProtectionGroup Text

The name of the protection group. You use this to identify the protection group in lists and to manage the protection group, for example to update, delete, or describe it.

protectionGroup_aggregation :: Lens' ProtectionGroup ProtectionGroupAggregation

Defines how Shield combines resource data for the group in order to detect, mitigate, and report events.

Sum - Use the total traffic across the group. This is a good choice for most cases. Examples include Elastic IP addresses for EC2 instances that scale manually or automatically.

Mean - Use the average of the traffic across the group. This is a good choice for resources that share traffic uniformly. Examples include accelerators and load balancers.

Max - Use the highest traffic from each resource. This is useful for resources that don't share traffic and for resources that share that traffic in a non-uniform way. Examples include Amazon CloudFront distributions and origin resources for CloudFront distributions.

protectionGroup_pattern :: Lens' ProtectionGroup ProtectionGroupPattern

The criteria to use to choose the protected resources for inclusion in the group. You can include all resources that have protections, provide a list of resource Amazon Resource Names (ARNs), or include all resources of a specified resource type.

protectionGroup_members :: Lens' ProtectionGroup [Text]

The Amazon Resource Names (ARNs) of the resources to include in the protection group. You must set this when you set Pattern to ARBITRARY and you must not set it for any other Pattern setting.

ProtectionGroupArbitraryPatternLimits

protectionGroupArbitraryPatternLimits_maxMembers :: Lens' ProtectionGroupArbitraryPatternLimits Integer

The maximum number of resources you can specify for a single arbitrary pattern in a protection group.

ProtectionGroupLimits

protectionGroupLimits_maxProtectionGroups :: Lens' ProtectionGroupLimits Integer

The maximum number of protection groups that you can have at one time.

protectionGroupLimits_patternTypeLimits :: Lens' ProtectionGroupLimits ProtectionGroupPatternTypeLimits

Limits settings by pattern type in the protection groups for your subscription.

ProtectionGroupPatternTypeLimits

protectionGroupPatternTypeLimits_arbitraryPatternLimits :: Lens' ProtectionGroupPatternTypeLimits ProtectionGroupArbitraryPatternLimits

Limits settings on protection groups with arbitrary pattern type.

ProtectionLimits

protectionLimits_protectedResourceTypeLimits :: Lens' ProtectionLimits [Limit]

The maximum number of resource types that you can specify in a protection.

SubResourceSummary

subResourceSummary_counters :: Lens' SubResourceSummary (Maybe [SummarizedCounter])

The counters that describe the details of the attack.

subResourceSummary_attackVectors :: Lens' SubResourceSummary (Maybe [SummarizedAttackVector])

The list of attack types and associated counters.

subResourceSummary_id :: Lens' SubResourceSummary (Maybe Text)

The unique identifier (ID) of the SubResource.

subResourceSummary_type :: Lens' SubResourceSummary (Maybe SubResourceType)

The SubResource type.

Subscription

subscription_timeCommitmentInSeconds :: Lens' Subscription (Maybe Natural)

The length, in seconds, of the Shield Advanced subscription for the account.

subscription_startTime :: Lens' Subscription (Maybe UTCTime)

The start time of the subscription, in Unix time in seconds. For more information see timestamp.

subscription_limits :: Lens' Subscription (Maybe [Limit])

Specifies how many protections of a given type you can create.

subscription_autoRenew :: Lens' Subscription (Maybe AutoRenew)

If ENABLED, the subscription will be automatically renewed at the end of the existing subscription period.

When you initally create a subscription, AutoRenew is set to ENABLED. You can change this by submitting an UpdateSubscription request. If the UpdateSubscription request does not included a value for AutoRenew, the existing value for AutoRenew remains unchanged.

subscription_endTime :: Lens' Subscription (Maybe UTCTime)

The date and time your subscription will end.

subscription_proactiveEngagementStatus :: Lens' Subscription (Maybe ProactiveEngagementStatus)

If ENABLED, the Shield Response Team (SRT) will use email and phone to notify contacts about escalations to the SRT and to initiate proactive customer support.

If PENDING, you have requested proactive engagement and the request is pending. The status changes to ENABLED when your request is fully processed.

If DISABLED, the SRT will not proactively notify contacts about escalations or to initiate proactive customer support.

subscription_subscriptionArn :: Lens' Subscription (Maybe Text)

The ARN (Amazon Resource Name) of the subscription.

subscription_subscriptionLimits :: Lens' Subscription SubscriptionLimits

Limits settings for your subscription.

SubscriptionLimits

subscriptionLimits_protectionLimits :: Lens' SubscriptionLimits ProtectionLimits

Limits settings on protections for your subscription.

subscriptionLimits_protectionGroupLimits :: Lens' SubscriptionLimits ProtectionGroupLimits

Limits settings on protection groups for your subscription.

SummarizedAttackVector

summarizedAttackVector_vectorCounters :: Lens' SummarizedAttackVector (Maybe [SummarizedCounter])

The list of counters that describe the details of the attack.

summarizedAttackVector_vectorType :: Lens' SummarizedAttackVector Text

The attack type, for example, SNMP reflection or SYN flood.

SummarizedCounter

summarizedCounter_max :: Lens' SummarizedCounter (Maybe Double)

The maximum value of the counter for a specified time period.

summarizedCounter_average :: Lens' SummarizedCounter (Maybe Double)

The average value of the counter for a specified time period.

summarizedCounter_n :: Lens' SummarizedCounter (Maybe Int)

The number of counters for a specified time period.

summarizedCounter_name :: Lens' SummarizedCounter (Maybe Text)

The counter name.

summarizedCounter_sum :: Lens' SummarizedCounter (Maybe Double)

The total of counter values for a specified time period.

summarizedCounter_unit :: Lens' SummarizedCounter (Maybe Text)

The unit of the counters.

Tag

tag_value :: Lens' Tag (Maybe Text)

Part of the key:value pair that defines a tag. You can use a tag value to describe a specific value within a category, such as "companyA" or "companyB." Tag values are case-sensitive.

tag_key :: Lens' Tag (Maybe Text)

Part of the key:value pair that defines a tag. You can use a tag key to describe a category of information, such as "customer." Tag keys are case-sensitive.

TimeRange

timeRange_fromInclusive :: Lens' TimeRange (Maybe UTCTime)

The start time, in Unix time in seconds. For more information see timestamp.

timeRange_toExclusive :: Lens' TimeRange (Maybe UTCTime)

The end time, in Unix time in seconds. For more information see timestamp.