Network.AWS.Shield.ListAttacks

Description

Returns all ongoing DDoS attacks or all DDoS attacks during a specified time period.

This operation returns paginated results.

Creating a Request

data ListAttacks

See: newListAttacks smart constructor.

Constructors

ListAttacks'

Fields

startTime :: Maybe TimeRange

The start of the time period for the attacks. This is a timestamp type. The sample request above indicates a number type because the default used by WAF is Unix time in seconds. However any valid timestamp format is allowed.

resourceArns :: Maybe [Text]

The ARN (Amazon Resource Name) of the resource that was attacked. If this is left blank, all applicable resources for this account will be included.

nextToken :: Maybe Text

The ListAttacksRequest.NextMarker value from a previous call to ListAttacksRequest. Pass null if this is the first call.

endTime :: Maybe TimeRange

The end of the time period for the attacks. This is a timestamp type. The sample request above indicates a number type because the default used by WAF is Unix time in seconds. However any valid timestamp format is allowed.

maxResults :: Maybe Natural

The maximum number of AttackSummary objects to return. If you leave this blank, Shield Advanced returns the first 20 results.

This is a maximum value. Shield Advanced might return the results in smaller batches. That is, the number of objects returned could be less than MaxResults, even if there are still more objects yet to return. If there are more objects to return, Shield Advanced returns a value in NextToken that you can use in your next request, to get the next batch of objects.

newListAttacks :: ListAttacks

Create a value of ListAttacks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTime:ListAttacks', listAttacks_startTime - The start of the time period for the attacks. This is a timestamp type. The sample request above indicates a number type because the default used by WAF is Unix time in seconds. However any valid timestamp format is allowed.

$sel:resourceArns:ListAttacks', listAttacks_resourceArns - The ARN (Amazon Resource Name) of the resource that was attacked. If this is left blank, all applicable resources for this account will be included.

$sel:nextToken:ListAttacks', listAttacks_nextToken - The ListAttacksRequest.NextMarker value from a previous call to ListAttacksRequest. Pass null if this is the first call.

$sel:endTime:ListAttacks', listAttacks_endTime - The end of the time period for the attacks. This is a timestamp type. The sample request above indicates a number type because the default used by WAF is Unix time in seconds. However any valid timestamp format is allowed.

$sel:maxResults:ListAttacks', listAttacks_maxResults - The maximum number of AttackSummary objects to return. If you leave this blank, Shield Advanced returns the first 20 results.

This is a maximum value. Shield Advanced might return the results in smaller batches. That is, the number of objects returned could be less than MaxResults, even if there are still more objects yet to return. If there are more objects to return, Shield Advanced returns a value in NextToken that you can use in your next request, to get the next batch of objects.

Request Lenses

listAttacks_startTime :: Lens' ListAttacks (Maybe TimeRange)

The start of the time period for the attacks. This is a timestamp type. The sample request above indicates a number type because the default used by WAF is Unix time in seconds. However any valid timestamp format is allowed.

listAttacks_resourceArns :: Lens' ListAttacks (Maybe [Text])

The ARN (Amazon Resource Name) of the resource that was attacked. If this is left blank, all applicable resources for this account will be included.

listAttacks_nextToken :: Lens' ListAttacks (Maybe Text)

The ListAttacksRequest.NextMarker value from a previous call to ListAttacksRequest. Pass null if this is the first call.

listAttacks_endTime :: Lens' ListAttacks (Maybe TimeRange)

The end of the time period for the attacks. This is a timestamp type. The sample request above indicates a number type because the default used by WAF is Unix time in seconds. However any valid timestamp format is allowed.

listAttacks_maxResults :: Lens' ListAttacks (Maybe Natural)

The maximum number of AttackSummary objects to return. If you leave this blank, Shield Advanced returns the first 20 results.

This is a maximum value. Shield Advanced might return the results in smaller batches. That is, the number of objects returned could be less than MaxResults, even if there are still more objects yet to return. If there are more objects to return, Shield Advanced returns a value in NextToken that you can use in your next request, to get the next batch of objects.

Destructuring the Response

data ListAttacksResponse

See: newListAttacksResponse smart constructor.

Constructors

ListAttacksResponse'

Fields

attackSummaries :: Maybe [AttackSummary]

The attack information for the specified time range.

nextToken :: Maybe Text

The token returned by a previous call to indicate that there is more data available. If not null, more results are available. Pass this value for the NextMarker parameter in a subsequent call to ListAttacks to retrieve the next set of items.

Shield Advanced might return the list of AttackSummary objects in batches smaller than the number specified by MaxResults. If there are more attack summary objects to return, Shield Advanced will always also return a NextToken.

httpStatus :: Int

The response's http status code.

newListAttacksResponse

Arguments

:: Int
$sel:httpStatus:ListAttacksResponse'

-> ListAttacksResponse

Create a value of ListAttacksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attackSummaries:ListAttacksResponse', listAttacksResponse_attackSummaries - The attack information for the specified time range.

$sel:nextToken:ListAttacks', listAttacksResponse_nextToken - The token returned by a previous call to indicate that there is more data available. If not null, more results are available. Pass this value for the NextMarker parameter in a subsequent call to ListAttacks to retrieve the next set of items.

Shield Advanced might return the list of AttackSummary objects in batches smaller than the number specified by MaxResults. If there are more attack summary objects to return, Shield Advanced will always also return a NextToken.

$sel:httpStatus:ListAttacksResponse', listAttacksResponse_httpStatus - The response's http status code.

Response Lenses

listAttacksResponse_attackSummaries :: Lens' ListAttacksResponse (Maybe [AttackSummary])

The attack information for the specified time range.

listAttacksResponse_nextToken :: Lens' ListAttacksResponse (Maybe Text)

The token returned by a previous call to indicate that there is more data available. If not null, more results are available. Pass this value for the NextMarker parameter in a subsequent call to ListAttacks to retrieve the next set of items.

Shield Advanced might return the list of AttackSummary objects in batches smaller than the number specified by MaxResults. If there are more attack summary objects to return, Shield Advanced will always also return a NextToken.

listAttacksResponse_httpStatus :: Lens' ListAttacksResponse Int

The response's http status code.