Network.AWS.Shield.Types

Description

Service Configuration

defaultService :: Service

API version 2016-06-02 of the Amazon Shield SDK configuration.

Errors

_OptimisticLockException :: AsError a => Getting (First ServiceError) a ServiceError

Exception that indicates that the resource state has been modified by another client. Retrieve the resource and then retry your request.

_InvalidResourceException :: AsError a => Getting (First ServiceError) a ServiceError

Exception that indicates that the resource is invalid. You might not have access to the resource, or the resource might not exist.

_InvalidPaginationTokenException :: AsError a => Getting (First ServiceError) a ServiceError

Exception that indicates that the NextToken specified in the request is invalid. Submit the request using the NextToken value that was returned in the response.

_InvalidOperationException :: AsError a => Getting (First ServiceError) a ServiceError

Exception that indicates that the operation would not cause any change to occur.

_ResourceAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

Exception indicating the specified resource already exists. If available, this exception includes details in additional properties.

_InternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

Exception that indicates that a problem occurred with the service infrastructure. You can retry the request.

_NoAssociatedRoleException :: AsError a => Getting (First ServiceError) a ServiceError

The ARN of the role that you specifed does not exist.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

Exception that indicates that the parameters passed to the API are invalid. If available, this exception includes details in additional properties.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

Exception that indicates the specified AttackId does not exist, or the requester does not have the appropriate permissions to access the AttackId.

_LockedSubscriptionException :: AsError a => Getting (First ServiceError) a ServiceError

You are trying to update a subscription that has not yet completed the 1-year commitment. You can change the AutoRenew parameter during the last 30 days of your subscription. This exception indicates that you are attempting to change AutoRenew prior to that period.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Exception indicating the specified resource does not exist. If available, this exception includes details in additional properties.

_AccessDeniedForDependencyException :: AsError a => Getting (First ServiceError) a ServiceError

In order to grant the necessary access to the Shield Response Team (SRT) the user submitting the request must have the iam:PassRole permission. This error indicates the user did not have the appropriate permissions. For more information, see Granting a User Permissions to Pass a Role to an Amazon Web Services Service.

_LimitsExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Exception that indicates that the operation would exceed a limit.

AttackLayer

newtype AttackLayer

Constructors

AttackLayer'

Fields

fromAttackLayer :: Text

Bundled Patterns

pattern AttackLayer_APPLICATION :: AttackLayer

pattern AttackLayer_NETWORK :: AttackLayer

AttackPropertyIdentifier

newtype AttackPropertyIdentifier

Constructors

AttackPropertyIdentifier'

Fields

fromAttackPropertyIdentifier :: Text

Bundled Patterns

pattern AttackPropertyIdentifier_DESTINATION_URL :: AttackPropertyIdentifier

pattern AttackPropertyIdentifier_REFERRER :: AttackPropertyIdentifier

pattern AttackPropertyIdentifier_SOURCE_ASN :: AttackPropertyIdentifier

pattern AttackPropertyIdentifier_SOURCE_COUNTRY :: AttackPropertyIdentifier

pattern AttackPropertyIdentifier_SOURCE_IP_ADDRESS :: AttackPropertyIdentifier

pattern AttackPropertyIdentifier_SOURCE_USER_AGENT :: AttackPropertyIdentifier

pattern AttackPropertyIdentifier_WORDPRESS_PINGBACK_REFLECTOR :: AttackPropertyIdentifier

pattern AttackPropertyIdentifier_WORDPRESS_PINGBACK_SOURCE :: AttackPropertyIdentifier

AutoRenew

newtype AutoRenew

Constructors

AutoRenew'

Fields

fromAutoRenew :: Text

Bundled Patterns

pattern AutoRenew_DISABLED :: AutoRenew

pattern AutoRenew_ENABLED :: AutoRenew

ProactiveEngagementStatus

newtype ProactiveEngagementStatus

Constructors

ProactiveEngagementStatus'

Fields

fromProactiveEngagementStatus :: Text

Bundled Patterns

pattern ProactiveEngagementStatus_DISABLED :: ProactiveEngagementStatus

pattern ProactiveEngagementStatus_ENABLED :: ProactiveEngagementStatus

pattern ProactiveEngagementStatus_PENDING :: ProactiveEngagementStatus

ProtectedResourceType

newtype ProtectedResourceType

Constructors

ProtectedResourceType'

Fields

fromProtectedResourceType :: Text

Bundled Patterns

pattern ProtectedResourceType_APPLICATION_LOAD_BALANCER :: ProtectedResourceType

pattern ProtectedResourceType_CLASSIC_LOAD_BALANCER :: ProtectedResourceType

pattern ProtectedResourceType_CLOUDFRONT_DISTRIBUTION :: ProtectedResourceType

pattern ProtectedResourceType_ELASTIC_IP_ALLOCATION :: ProtectedResourceType

pattern ProtectedResourceType_GLOBAL_ACCELERATOR :: ProtectedResourceType

pattern ProtectedResourceType_ROUTE_53_HOSTED_ZONE :: ProtectedResourceType

ProtectionGroupAggregation

newtype ProtectionGroupAggregation

Constructors

ProtectionGroupAggregation'

Fields

fromProtectionGroupAggregation :: Text

Bundled Patterns

pattern ProtectionGroupAggregation_MAX :: ProtectionGroupAggregation

pattern ProtectionGroupAggregation_MEAN :: ProtectionGroupAggregation

pattern ProtectionGroupAggregation_SUM :: ProtectionGroupAggregation

ProtectionGroupPattern

newtype ProtectionGroupPattern

Constructors

ProtectionGroupPattern'

Fields

fromProtectionGroupPattern :: Text

Bundled Patterns

pattern ProtectionGroupPattern_ALL :: ProtectionGroupPattern

pattern ProtectionGroupPattern_ARBITRARY :: ProtectionGroupPattern

pattern ProtectionGroupPattern_BY_RESOURCE_TYPE :: ProtectionGroupPattern

SubResourceType

newtype SubResourceType

Constructors

SubResourceType'

Fields

fromSubResourceType :: Text

Bundled Patterns

pattern SubResourceType_IP :: SubResourceType

pattern SubResourceType_URL :: SubResourceType

SubscriptionState

newtype SubscriptionState

Constructors

SubscriptionState'

Fields

fromSubscriptionState :: Text

Bundled Patterns

pattern SubscriptionState_ACTIVE :: SubscriptionState

pattern SubscriptionState_INACTIVE :: SubscriptionState

Unit

newtype Unit

Constructors

Unit'

Fields

fromUnit :: Text

Bundled Patterns

pattern Unit_BITS :: Unit

pattern Unit_BYTES :: Unit

pattern Unit_PACKETS :: Unit

pattern Unit_REQUESTS :: Unit

AttackDetail

data AttackDetail

The details of a DDoS attack.

See: newAttackDetail smart constructor.

Constructors

AttackDetail'

Fields

resourceArn :: Maybe Text

The ARN (Amazon Resource Name) of the resource that was attacked.

attackProperties :: Maybe [AttackProperty]

The array of objects that provide details of the Shield event.

For infrastructure layer events (L3 and L4 events) after January 25, 2021, you can view metrics for top contributors in Amazon CloudWatch metrics. For more information, see Shield metrics and alarms in the WAF Developer Guide.

startTime :: Maybe POSIX

The time the attack started, in Unix time in seconds. For more information see timestamp.

endTime :: Maybe POSIX

The time the attack ended, in Unix time in seconds. For more information see timestamp.

attackCounters :: Maybe [SummarizedCounter]

List of counters that describe the attack for the specified time period.

mitigations :: Maybe [Mitigation]

List of mitigation actions taken for the attack.

attackId :: Maybe Text

The unique identifier (ID) of the attack.

subResources :: Maybe [SubResourceSummary]

If applicable, additional detail about the resource being attacked, for example, IP address or URL.

newAttackDetail :: AttackDetail

Create a value of AttackDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:AttackDetail', attackDetail_resourceArn - The ARN (Amazon Resource Name) of the resource that was attacked.

$sel:attackProperties:AttackDetail', attackDetail_attackProperties - The array of objects that provide details of the Shield event.

For infrastructure layer events (L3 and L4 events) after January 25, 2021, you can view metrics for top contributors in Amazon CloudWatch metrics. For more information, see Shield metrics and alarms in the WAF Developer Guide.

$sel:startTime:AttackDetail', attackDetail_startTime - The time the attack started, in Unix time in seconds. For more information see timestamp.

$sel:endTime:AttackDetail', attackDetail_endTime - The time the attack ended, in Unix time in seconds. For more information see timestamp.

$sel:attackCounters:AttackDetail', attackDetail_attackCounters - List of counters that describe the attack for the specified time period.

$sel:mitigations:AttackDetail', attackDetail_mitigations - List of mitigation actions taken for the attack.

$sel:attackId:AttackDetail', attackDetail_attackId - The unique identifier (ID) of the attack.

$sel:subResources:AttackDetail', attackDetail_subResources - If applicable, additional detail about the resource being attacked, for example, IP address or URL.

attackDetail_resourceArn :: Lens' AttackDetail (Maybe Text)

The ARN (Amazon Resource Name) of the resource that was attacked.

attackDetail_attackProperties :: Lens' AttackDetail (Maybe [AttackProperty])

The array of objects that provide details of the Shield event.

For infrastructure layer events (L3 and L4 events) after January 25, 2021, you can view metrics for top contributors in Amazon CloudWatch metrics. For more information, see Shield metrics and alarms in the WAF Developer Guide.

attackDetail_startTime :: Lens' AttackDetail (Maybe UTCTime)

The time the attack started, in Unix time in seconds. For more information see timestamp.

attackDetail_endTime :: Lens' AttackDetail (Maybe UTCTime)

The time the attack ended, in Unix time in seconds. For more information see timestamp.

attackDetail_attackCounters :: Lens' AttackDetail (Maybe [SummarizedCounter])

List of counters that describe the attack for the specified time period.

attackDetail_mitigations :: Lens' AttackDetail (Maybe [Mitigation])

List of mitigation actions taken for the attack.

attackDetail_attackId :: Lens' AttackDetail (Maybe Text)

The unique identifier (ID) of the attack.

attackDetail_subResources :: Lens' AttackDetail (Maybe [SubResourceSummary])

If applicable, additional detail about the resource being attacked, for example, IP address or URL.

AttackProperty

data AttackProperty

Details of a Shield event. This is provided as part of an AttackDetail.

See: newAttackProperty smart constructor.

Constructors

AttackProperty'

Fields

unit :: Maybe Unit

The unit used for the Contributor Value property.

total :: Maybe Integer

The total contributions made to this Shield event by all contributors.

attackPropertyIdentifier :: Maybe AttackPropertyIdentifier

Defines the Shield event property information that is provided. The WORDPRESS_PINGBACK_REFLECTOR and WORDPRESS_PINGBACK_SOURCE values are valid only for WordPress reflective pingback events.

topContributors :: Maybe [Contributor]

Contributor objects for the top five contributors to a Shield event.

attackLayer :: Maybe AttackLayer

The type of Shield event that was observed. NETWORK indicates layer 3 and layer 4 events and APPLICATION indicates layer 7 events.

For infrastructure layer events (L3 and L4 events) after January 25, 2021, you can view metrics for top contributors in Amazon CloudWatch metrics. For more information, see Shield metrics and alarms in the WAF Developer Guide.

newAttackProperty :: AttackProperty

Create a value of AttackProperty with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unit:AttackProperty', attackProperty_unit - The unit used for the Contributor Value property.

$sel:total:AttackProperty', attackProperty_total - The total contributions made to this Shield event by all contributors.

$sel:attackPropertyIdentifier:AttackProperty', attackProperty_attackPropertyIdentifier - Defines the Shield event property information that is provided. The WORDPRESS_PINGBACK_REFLECTOR and WORDPRESS_PINGBACK_SOURCE values are valid only for WordPress reflective pingback events.

$sel:topContributors:AttackProperty', attackProperty_topContributors - Contributor objects for the top five contributors to a Shield event.

$sel:attackLayer:AttackProperty', attackProperty_attackLayer - The type of Shield event that was observed. NETWORK indicates layer 3 and layer 4 events and APPLICATION indicates layer 7 events.

For infrastructure layer events (L3 and L4 events) after January 25, 2021, you can view metrics for top contributors in Amazon CloudWatch metrics. For more information, see Shield metrics and alarms in the WAF Developer Guide.

attackProperty_unit :: Lens' AttackProperty (Maybe Unit)

The unit used for the Contributor Value property.

attackProperty_total :: Lens' AttackProperty (Maybe Integer)

The total contributions made to this Shield event by all contributors.

attackProperty_attackPropertyIdentifier :: Lens' AttackProperty (Maybe AttackPropertyIdentifier)

Defines the Shield event property information that is provided. The WORDPRESS_PINGBACK_REFLECTOR and WORDPRESS_PINGBACK_SOURCE values are valid only for WordPress reflective pingback events.

attackProperty_topContributors :: Lens' AttackProperty (Maybe [Contributor])

Contributor objects for the top five contributors to a Shield event.

attackProperty_attackLayer :: Lens' AttackProperty (Maybe AttackLayer)

The type of Shield event that was observed. NETWORK indicates layer 3 and layer 4 events and APPLICATION indicates layer 7 events.

For infrastructure layer events (L3 and L4 events) after January 25, 2021, you can view metrics for top contributors in Amazon CloudWatch metrics. For more information, see Shield metrics and alarms in the WAF Developer Guide.

AttackStatisticsDataItem

data AttackStatisticsDataItem

A single attack statistics data record. This is returned by DescribeAttackStatistics along with a time range indicating the time period that the attack statistics apply to.

See: newAttackStatisticsDataItem smart constructor.

Constructors

AttackStatisticsDataItem'

Fields

attackVolume :: Maybe AttackVolume

Information about the volume of attacks during the time period. If the accompanying AttackCount is zero, this setting might be empty.

attackCount :: Integer

The number of attacks detected during the time period. This is always present, but might be zero.

newAttackStatisticsDataItem

Arguments

:: Integer
$sel:attackCount:AttackStatisticsDataItem'

-> AttackStatisticsDataItem

Create a value of AttackStatisticsDataItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attackVolume:AttackStatisticsDataItem', attackStatisticsDataItem_attackVolume - Information about the volume of attacks during the time period. If the accompanying AttackCount is zero, this setting might be empty.

$sel:attackCount:AttackStatisticsDataItem', attackStatisticsDataItem_attackCount - The number of attacks detected during the time period. This is always present, but might be zero.

attackStatisticsDataItem_attackVolume :: Lens' AttackStatisticsDataItem (Maybe AttackVolume)

Information about the volume of attacks during the time period. If the accompanying AttackCount is zero, this setting might be empty.

attackStatisticsDataItem_attackCount :: Lens' AttackStatisticsDataItem Integer

The number of attacks detected during the time period. This is always present, but might be zero.

AttackSummary

data AttackSummary

Summarizes all DDoS attacks for a specified time period.

See: newAttackSummary smart constructor.

Constructors

AttackSummary'

Fields

resourceArn :: Maybe Text

The ARN (Amazon Resource Name) of the resource that was attacked.

startTime :: Maybe POSIX

The start time of the attack, in Unix time in seconds. For more information see timestamp.

endTime :: Maybe POSIX

The end time of the attack, in Unix time in seconds. For more information see timestamp.

attackId :: Maybe Text

The unique identifier (ID) of the attack.

attackVectors :: Maybe [AttackVectorDescription]

The list of attacks for a specified time period.

newAttackSummary :: AttackSummary

Create a value of AttackSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:AttackSummary', attackSummary_resourceArn - The ARN (Amazon Resource Name) of the resource that was attacked.

$sel:startTime:AttackSummary', attackSummary_startTime - The start time of the attack, in Unix time in seconds. For more information see timestamp.

$sel:endTime:AttackSummary', attackSummary_endTime - The end time of the attack, in Unix time in seconds. For more information see timestamp.

$sel:attackId:AttackSummary', attackSummary_attackId - The unique identifier (ID) of the attack.

$sel:attackVectors:AttackSummary', attackSummary_attackVectors - The list of attacks for a specified time period.

attackSummary_resourceArn :: Lens' AttackSummary (Maybe Text)

The ARN (Amazon Resource Name) of the resource that was attacked.

attackSummary_startTime :: Lens' AttackSummary (Maybe UTCTime)

The start time of the attack, in Unix time in seconds. For more information see timestamp.

attackSummary_endTime :: Lens' AttackSummary (Maybe UTCTime)

The end time of the attack, in Unix time in seconds. For more information see timestamp.

attackSummary_attackId :: Lens' AttackSummary (Maybe Text)

The unique identifier (ID) of the attack.

attackSummary_attackVectors :: Lens' AttackSummary (Maybe [AttackVectorDescription])

The list of attacks for a specified time period.

AttackVectorDescription

data AttackVectorDescription

Describes the attack.

See: newAttackVectorDescription smart constructor.

Constructors

AttackVectorDescription'

Fields

vectorType :: Text

The attack type. Valid values:

UDP_TRAFFIC

UDP_FRAGMENT

GENERIC_UDP_REFLECTION

DNS_REFLECTION

NTP_REFLECTION

CHARGEN_REFLECTION

SSDP_REFLECTION

PORT_MAPPER

RIP_REFLECTION

SNMP_REFLECTION

MSSQL_REFLECTION

NET_BIOS_REFLECTION

SYN_FLOOD

ACK_FLOOD

REQUEST_FLOOD

HTTP_REFLECTION

UDS_REFLECTION

MEMCACHED_REFLECTION

newAttackVectorDescription

Arguments

:: Text
$sel:vectorType:AttackVectorDescription'

-> AttackVectorDescription

Create a value of AttackVectorDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vectorType:AttackVectorDescription', attackVectorDescription_vectorType - The attack type. Valid values:

UDP_TRAFFIC

UDP_FRAGMENT

GENERIC_UDP_REFLECTION

DNS_REFLECTION

NTP_REFLECTION

CHARGEN_REFLECTION

SSDP_REFLECTION

PORT_MAPPER

RIP_REFLECTION

SNMP_REFLECTION

MSSQL_REFLECTION

NET_BIOS_REFLECTION

SYN_FLOOD

ACK_FLOOD

REQUEST_FLOOD

HTTP_REFLECTION

UDS_REFLECTION

MEMCACHED_REFLECTION

attackVectorDescription_vectorType :: Lens' AttackVectorDescription Text

The attack type. Valid values:

UDP_TRAFFIC

UDP_FRAGMENT

GENERIC_UDP_REFLECTION

DNS_REFLECTION

NTP_REFLECTION

CHARGEN_REFLECTION

SSDP_REFLECTION

PORT_MAPPER

RIP_REFLECTION

SNMP_REFLECTION

MSSQL_REFLECTION

NET_BIOS_REFLECTION

SYN_FLOOD

ACK_FLOOD

REQUEST_FLOOD

HTTP_REFLECTION

UDS_REFLECTION

MEMCACHED_REFLECTION

AttackVolume

data AttackVolume

Information about the volume of attacks during the time period, included in an AttackStatisticsDataItem. If the accompanying AttackCount in the statistics object is zero, this setting might be empty.

See: newAttackVolume smart constructor.

Constructors

AttackVolume'

Fields

packetsPerSecond :: Maybe AttackVolumeStatistics

A statistics object that uses packets per second as the unit. This is included for network level attacks.

bitsPerSecond :: Maybe AttackVolumeStatistics

A statistics object that uses bits per second as the unit. This is included for network level attacks.

requestsPerSecond :: Maybe AttackVolumeStatistics

A statistics object that uses requests per second as the unit. This is included for application level attacks, and is only available for accounts that are subscribed to Shield Advanced.

newAttackVolume :: AttackVolume

Create a value of AttackVolume with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packetsPerSecond:AttackVolume', attackVolume_packetsPerSecond - A statistics object that uses packets per second as the unit. This is included for network level attacks.

$sel:bitsPerSecond:AttackVolume', attackVolume_bitsPerSecond - A statistics object that uses bits per second as the unit. This is included for network level attacks.

$sel:requestsPerSecond:AttackVolume', attackVolume_requestsPerSecond - A statistics object that uses requests per second as the unit. This is included for application level attacks, and is only available for accounts that are subscribed to Shield Advanced.

attackVolume_packetsPerSecond :: Lens' AttackVolume (Maybe AttackVolumeStatistics)

A statistics object that uses packets per second as the unit. This is included for network level attacks.

attackVolume_bitsPerSecond :: Lens' AttackVolume (Maybe AttackVolumeStatistics)

A statistics object that uses bits per second as the unit. This is included for network level attacks.

attackVolume_requestsPerSecond :: Lens' AttackVolume (Maybe AttackVolumeStatistics)

A statistics object that uses requests per second as the unit. This is included for application level attacks, and is only available for accounts that are subscribed to Shield Advanced.

AttackVolumeStatistics

data AttackVolumeStatistics

Statistics objects for the various data types in AttackVolume.

See: newAttackVolumeStatistics smart constructor.

Constructors

AttackVolumeStatistics'

Fields

max :: Double

The maximum attack volume observed for the given unit.

newAttackVolumeStatistics

Arguments

:: Double
$sel:max:AttackVolumeStatistics'

-> AttackVolumeStatistics

Create a value of AttackVolumeStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:max:AttackVolumeStatistics', attackVolumeStatistics_max - The maximum attack volume observed for the given unit.

attackVolumeStatistics_max :: Lens' AttackVolumeStatistics Double

The maximum attack volume observed for the given unit.

Contributor

data Contributor

A contributor to the attack and their contribution.

See: newContributor smart constructor.

Constructors

Contributor'

Fields

name :: Maybe Text

The name of the contributor. This is dependent on the AttackPropertyIdentifier. For example, if the AttackPropertyIdentifier is SOURCE_COUNTRY, the Name could be United States.

value :: Maybe Integer

The contribution of this contributor expressed in Protection units. For example 10,000.

newContributor :: Contributor

Create a value of Contributor with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:Contributor', contributor_name - The name of the contributor. This is dependent on the AttackPropertyIdentifier. For example, if the AttackPropertyIdentifier is SOURCE_COUNTRY, the Name could be United States.

$sel:value:Contributor', contributor_value - The contribution of this contributor expressed in Protection units. For example 10,000.

contributor_name :: Lens' Contributor (Maybe Text)

The name of the contributor. This is dependent on the AttackPropertyIdentifier. For example, if the AttackPropertyIdentifier is SOURCE_COUNTRY, the Name could be United States.

contributor_value :: Lens' Contributor (Maybe Integer)

The contribution of this contributor expressed in Protection units. For example 10,000.

EmergencyContact

data EmergencyContact

Contact information that the SRT can use to contact you if you have proactive engagement enabled, for escalations to the SRT and to initiate proactive customer support.

See: newEmergencyContact smart constructor.

Constructors

EmergencyContact'

Fields

phoneNumber :: Maybe Text

The phone number for the contact.

contactNotes :: Maybe Text

Additional notes regarding the contact.

emailAddress :: Text

The email address for the contact.

newEmergencyContact

Arguments

:: Text
$sel:emailAddress:EmergencyContact'

-> EmergencyContact

Create a value of EmergencyContact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneNumber:EmergencyContact', emergencyContact_phoneNumber - The phone number for the contact.

$sel:contactNotes:EmergencyContact', emergencyContact_contactNotes - Additional notes regarding the contact.

$sel:emailAddress:EmergencyContact', emergencyContact_emailAddress - The email address for the contact.

emergencyContact_phoneNumber :: Lens' EmergencyContact (Maybe Text)

The phone number for the contact.

emergencyContact_contactNotes :: Lens' EmergencyContact (Maybe Text)

Additional notes regarding the contact.

emergencyContact_emailAddress :: Lens' EmergencyContact Text

The email address for the contact.

Limit

data Limit

Specifies how many protections of a given type you can create.

See: newLimit smart constructor.

Constructors

Limit'

Fields

max :: Maybe Integer

The maximum number of protections that can be created for the specified Type.

type' :: Maybe Text

The type of protection.

newLimit :: Limit

Create a value of Limit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:max:Limit', limit_max - The maximum number of protections that can be created for the specified Type.

$sel:type':Limit', limit_type - The type of protection.

limit_max :: Lens' Limit (Maybe Integer)

The maximum number of protections that can be created for the specified Type.

limit_type :: Lens' Limit (Maybe Text)

The type of protection.

Mitigation

data Mitigation

The mitigation applied to a DDoS attack.

See: newMitigation smart constructor.

Constructors

Mitigation'

Fields

mitigationName :: Maybe Text

The name of the mitigation taken for this attack.

newMitigation :: Mitigation

Create a value of Mitigation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mitigationName:Mitigation', mitigation_mitigationName - The name of the mitigation taken for this attack.

mitigation_mitigationName :: Lens' Mitigation (Maybe Text)

The name of the mitigation taken for this attack.

Protection

data Protection

An object that represents a resource that is under DDoS protection.

See: newProtection smart constructor.

Constructors

Protection'

Fields

resourceArn :: Maybe Text

The ARN (Amazon Resource Name) of the Amazon Web Services resource that is protected.

protectionArn :: Maybe Text

The ARN (Amazon Resource Name) of the protection.

id :: Maybe Text

The unique identifier (ID) of the protection.

name :: Maybe Text

The name of the protection. For example, My CloudFront distributions.

healthCheckIds :: Maybe [Text]

The unique identifier (ID) for the Route 53 health check that's associated with the protection.

newProtection :: Protection

Create a value of Protection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:Protection', protection_resourceArn - The ARN (Amazon Resource Name) of the Amazon Web Services resource that is protected.

$sel:protectionArn:Protection', protection_protectionArn - The ARN (Amazon Resource Name) of the protection.

$sel:id:Protection', protection_id - The unique identifier (ID) of the protection.

$sel:name:Protection', protection_name - The name of the protection. For example, My CloudFront distributions.

$sel:healthCheckIds:Protection', protection_healthCheckIds - The unique identifier (ID) for the Route 53 health check that's associated with the protection.

protection_resourceArn :: Lens' Protection (Maybe Text)

The ARN (Amazon Resource Name) of the Amazon Web Services resource that is protected.

protection_protectionArn :: Lens' Protection (Maybe Text)

The ARN (Amazon Resource Name) of the protection.

protection_id :: Lens' Protection (Maybe Text)

The unique identifier (ID) of the protection.

protection_name :: Lens' Protection (Maybe Text)

The name of the protection. For example, My CloudFront distributions.

protection_healthCheckIds :: Lens' Protection (Maybe [Text])

The unique identifier (ID) for the Route 53 health check that's associated with the protection.

ProtectionGroup

data ProtectionGroup

A grouping of protected resources that you and Shield Advanced can monitor as a collective. This resource grouping improves the accuracy of detection and reduces false positives.

See: newProtectionGroup smart constructor.

Constructors

ProtectionGroup'

Fields

resourceType :: Maybe ProtectedResourceType

The resource type to include in the protection group. All protected resources of this type are included in the protection group. You must set this when you set Pattern to BY_RESOURCE_TYPE and you must not set it for any other Pattern setting.

protectionGroupArn :: Maybe Text

The ARN (Amazon Resource Name) of the protection group.

protectionGroupId :: Text

The name of the protection group. You use this to identify the protection group in lists and to manage the protection group, for example to update, delete, or describe it.

aggregation :: ProtectionGroupAggregation

Defines how Shield combines resource data for the group in order to detect, mitigate, and report events.

Sum - Use the total traffic across the group. This is a good choice for most cases. Examples include Elastic IP addresses for EC2 instances that scale manually or automatically.

Mean - Use the average of the traffic across the group. This is a good choice for resources that share traffic uniformly. Examples include accelerators and load balancers.

Max - Use the highest traffic from each resource. This is useful for resources that don't share traffic and for resources that share that traffic in a non-uniform way. Examples include Amazon CloudFront distributions and origin resources for CloudFront distributions.

pattern' :: ProtectionGroupPattern

The criteria to use to choose the protected resources for inclusion in the group. You can include all resources that have protections, provide a list of resource Amazon Resource Names (ARNs), or include all resources of a specified resource type.

members :: [Text]

The Amazon Resource Names (ARNs) of the resources to include in the protection group. You must set this when you set Pattern to ARBITRARY and you must not set it for any other Pattern setting.

newProtectionGroup

Arguments

:: Text
$sel:protectionGroupId:ProtectionGroup'

-> ProtectionGroupAggregation
$sel:aggregation:ProtectionGroup'

-> ProtectionGroupPattern
$sel:pattern':ProtectionGroup'

-> ProtectionGroup

Create a value of ProtectionGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ProtectionGroup', protectionGroup_resourceType - The resource type to include in the protection group. All protected resources of this type are included in the protection group. You must set this when you set Pattern to BY_RESOURCE_TYPE and you must not set it for any other Pattern setting.

$sel:protectionGroupArn:ProtectionGroup', protectionGroup_protectionGroupArn - The ARN (Amazon Resource Name) of the protection group.

$sel:protectionGroupId:ProtectionGroup', protectionGroup_protectionGroupId - The name of the protection group. You use this to identify the protection group in lists and to manage the protection group, for example to update, delete, or describe it.

$sel:aggregation:ProtectionGroup', protectionGroup_aggregation - Defines how Shield combines resource data for the group in order to detect, mitigate, and report events.

Sum - Use the total traffic across the group. This is a good choice for most cases. Examples include Elastic IP addresses for EC2 instances that scale manually or automatically.

Mean - Use the average of the traffic across the group. This is a good choice for resources that share traffic uniformly. Examples include accelerators and load balancers.

Max - Use the highest traffic from each resource. This is useful for resources that don't share traffic and for resources that share that traffic in a non-uniform way. Examples include Amazon CloudFront distributions and origin resources for CloudFront distributions.

$sel:pattern':ProtectionGroup', protectionGroup_pattern - The criteria to use to choose the protected resources for inclusion in the group. You can include all resources that have protections, provide a list of resource Amazon Resource Names (ARNs), or include all resources of a specified resource type.

$sel:members:ProtectionGroup', protectionGroup_members - The Amazon Resource Names (ARNs) of the resources to include in the protection group. You must set this when you set Pattern to ARBITRARY and you must not set it for any other Pattern setting.

protectionGroup_resourceType :: Lens' ProtectionGroup (Maybe ProtectedResourceType)

The resource type to include in the protection group. All protected resources of this type are included in the protection group. You must set this when you set Pattern to BY_RESOURCE_TYPE and you must not set it for any other Pattern setting.

protectionGroup_protectionGroupArn :: Lens' ProtectionGroup (Maybe Text)

The ARN (Amazon Resource Name) of the protection group.

protectionGroup_protectionGroupId :: Lens' ProtectionGroup Text

The name of the protection group. You use this to identify the protection group in lists and to manage the protection group, for example to update, delete, or describe it.

protectionGroup_aggregation :: Lens' ProtectionGroup ProtectionGroupAggregation

Defines how Shield combines resource data for the group in order to detect, mitigate, and report events.

Sum - Use the total traffic across the group. This is a good choice for most cases. Examples include Elastic IP addresses for EC2 instances that scale manually or automatically.

Mean - Use the average of the traffic across the group. This is a good choice for resources that share traffic uniformly. Examples include accelerators and load balancers.

Max - Use the highest traffic from each resource. This is useful for resources that don't share traffic and for resources that share that traffic in a non-uniform way. Examples include Amazon CloudFront distributions and origin resources for CloudFront distributions.

protectionGroup_pattern :: Lens' ProtectionGroup ProtectionGroupPattern

The criteria to use to choose the protected resources for inclusion in the group. You can include all resources that have protections, provide a list of resource Amazon Resource Names (ARNs), or include all resources of a specified resource type.

protectionGroup_members :: Lens' ProtectionGroup [Text]

The Amazon Resource Names (ARNs) of the resources to include in the protection group. You must set this when you set Pattern to ARBITRARY and you must not set it for any other Pattern setting.

ProtectionGroupArbitraryPatternLimits

data ProtectionGroupArbitraryPatternLimits

Limits settings on protection groups with arbitrary pattern type.

See: newProtectionGroupArbitraryPatternLimits smart constructor.

Constructors

ProtectionGroupArbitraryPatternLimits'

Fields

maxMembers :: Integer

The maximum number of resources you can specify for a single arbitrary pattern in a protection group.

newProtectionGroupArbitraryPatternLimits

Arguments

:: Integer
$sel:maxMembers:ProtectionGroupArbitraryPatternLimits'

-> ProtectionGroupArbitraryPatternLimits

Create a value of ProtectionGroupArbitraryPatternLimits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxMembers:ProtectionGroupArbitraryPatternLimits', protectionGroupArbitraryPatternLimits_maxMembers - The maximum number of resources you can specify for a single arbitrary pattern in a protection group.

protectionGroupArbitraryPatternLimits_maxMembers :: Lens' ProtectionGroupArbitraryPatternLimits Integer

The maximum number of resources you can specify for a single arbitrary pattern in a protection group.

ProtectionGroupLimits

data ProtectionGroupLimits

Limits settings on protection groups for your subscription.

See: newProtectionGroupLimits smart constructor.

Constructors

ProtectionGroupLimits'

Fields

maxProtectionGroups :: Integer

The maximum number of protection groups that you can have at one time.

patternTypeLimits :: ProtectionGroupPatternTypeLimits

Limits settings by pattern type in the protection groups for your subscription.

newProtectionGroupLimits

Arguments

:: Integer
$sel:maxProtectionGroups:ProtectionGroupLimits'

-> ProtectionGroupPatternTypeLimits
$sel:patternTypeLimits:ProtectionGroupLimits'

-> ProtectionGroupLimits

Create a value of ProtectionGroupLimits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxProtectionGroups:ProtectionGroupLimits', protectionGroupLimits_maxProtectionGroups - The maximum number of protection groups that you can have at one time.

$sel:patternTypeLimits:ProtectionGroupLimits', protectionGroupLimits_patternTypeLimits - Limits settings by pattern type in the protection groups for your subscription.

protectionGroupLimits_maxProtectionGroups :: Lens' ProtectionGroupLimits Integer

The maximum number of protection groups that you can have at one time.

protectionGroupLimits_patternTypeLimits :: Lens' ProtectionGroupLimits ProtectionGroupPatternTypeLimits

Limits settings by pattern type in the protection groups for your subscription.

ProtectionGroupPatternTypeLimits

data ProtectionGroupPatternTypeLimits

Limits settings by pattern type in the protection groups for your subscription.

See: newProtectionGroupPatternTypeLimits smart constructor.

Constructors

ProtectionGroupPatternTypeLimits'

Fields

arbitraryPatternLimits :: ProtectionGroupArbitraryPatternLimits

Limits settings on protection groups with arbitrary pattern type.

newProtectionGroupPatternTypeLimits

Arguments

:: ProtectionGroupArbitraryPatternLimits
$sel:arbitraryPatternLimits:ProtectionGroupPatternTypeLimits'

-> ProtectionGroupPatternTypeLimits

Create a value of ProtectionGroupPatternTypeLimits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arbitraryPatternLimits:ProtectionGroupPatternTypeLimits', protectionGroupPatternTypeLimits_arbitraryPatternLimits - Limits settings on protection groups with arbitrary pattern type.

protectionGroupPatternTypeLimits_arbitraryPatternLimits :: Lens' ProtectionGroupPatternTypeLimits ProtectionGroupArbitraryPatternLimits

Limits settings on protection groups with arbitrary pattern type.

ProtectionLimits

data ProtectionLimits

Limits settings on protections for your subscription.

See: newProtectionLimits smart constructor.

Constructors

ProtectionLimits'

Fields

protectedResourceTypeLimits :: [Limit]

The maximum number of resource types that you can specify in a protection.

newProtectionLimits :: ProtectionLimits

Create a value of ProtectionLimits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protectedResourceTypeLimits:ProtectionLimits', protectionLimits_protectedResourceTypeLimits - The maximum number of resource types that you can specify in a protection.

protectionLimits_protectedResourceTypeLimits :: Lens' ProtectionLimits [Limit]

The maximum number of resource types that you can specify in a protection.

SubResourceSummary

data SubResourceSummary

The attack information for the specified SubResource.

See: newSubResourceSummary smart constructor.

Constructors

SubResourceSummary'

Fields

counters :: Maybe [SummarizedCounter]

The counters that describe the details of the attack.

id :: Maybe Text

The unique identifier (ID) of the SubResource.

type' :: Maybe SubResourceType

The SubResource type.

attackVectors :: Maybe [SummarizedAttackVector]

The list of attack types and associated counters.

newSubResourceSummary :: SubResourceSummary

Create a value of SubResourceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:counters:SubResourceSummary', subResourceSummary_counters - The counters that describe the details of the attack.

$sel:id:SubResourceSummary', subResourceSummary_id - The unique identifier (ID) of the SubResource.

$sel:type':SubResourceSummary', subResourceSummary_type - The SubResource type.

$sel:attackVectors:SubResourceSummary', subResourceSummary_attackVectors - The list of attack types and associated counters.

subResourceSummary_counters :: Lens' SubResourceSummary (Maybe [SummarizedCounter])

The counters that describe the details of the attack.

subResourceSummary_id :: Lens' SubResourceSummary (Maybe Text)

The unique identifier (ID) of the SubResource.

subResourceSummary_type :: Lens' SubResourceSummary (Maybe SubResourceType)

The SubResource type.

subResourceSummary_attackVectors :: Lens' SubResourceSummary (Maybe [SummarizedAttackVector])

The list of attack types and associated counters.

Subscription

data Subscription

Information about the Shield Advanced subscription for an account.

See: newSubscription smart constructor.

Constructors

Subscription'

Fields

autoRenew :: Maybe AutoRenew

If ENABLED, the subscription will be automatically renewed at the end of the existing subscription period.

When you initally create a subscription, AutoRenew is set to ENABLED. You can change this by submitting an UpdateSubscription request. If the UpdateSubscription request does not included a value for AutoRenew, the existing value for AutoRenew remains unchanged.

proactiveEngagementStatus :: Maybe ProactiveEngagementStatus

If ENABLED, the Shield Response Team (SRT) will use email and phone to notify contacts about escalations to the SRT and to initiate proactive customer support.

If PENDING, you have requested proactive engagement and the request is pending. The status changes to ENABLED when your request is fully processed.

If DISABLED, the SRT will not proactively notify contacts about escalations or to initiate proactive customer support.

startTime :: Maybe POSIX

The start time of the subscription, in Unix time in seconds. For more information see timestamp.

endTime :: Maybe POSIX

The date and time your subscription will end.

limits :: Maybe [Limit]

Specifies how many protections of a given type you can create.

subscriptionArn :: Maybe Text

The ARN (Amazon Resource Name) of the subscription.

timeCommitmentInSeconds :: Maybe Natural

The length, in seconds, of the Shield Advanced subscription for the account.

subscriptionLimits :: SubscriptionLimits

Limits settings for your subscription.

newSubscription

Arguments

:: SubscriptionLimits
$sel:subscriptionLimits:Subscription'

-> Subscription

Create a value of Subscription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoRenew:Subscription', subscription_autoRenew - If ENABLED, the subscription will be automatically renewed at the end of the existing subscription period.

When you initally create a subscription, AutoRenew is set to ENABLED. You can change this by submitting an UpdateSubscription request. If the UpdateSubscription request does not included a value for AutoRenew, the existing value for AutoRenew remains unchanged.

$sel:proactiveEngagementStatus:Subscription', subscription_proactiveEngagementStatus - If ENABLED, the Shield Response Team (SRT) will use email and phone to notify contacts about escalations to the SRT and to initiate proactive customer support.

If PENDING, you have requested proactive engagement and the request is pending. The status changes to ENABLED when your request is fully processed.

If DISABLED, the SRT will not proactively notify contacts about escalations or to initiate proactive customer support.

$sel:startTime:Subscription', subscription_startTime - The start time of the subscription, in Unix time in seconds. For more information see timestamp.

$sel:endTime:Subscription', subscription_endTime - The date and time your subscription will end.

$sel:limits:Subscription', subscription_limits - Specifies how many protections of a given type you can create.

$sel:subscriptionArn:Subscription', subscription_subscriptionArn - The ARN (Amazon Resource Name) of the subscription.

$sel:timeCommitmentInSeconds:Subscription', subscription_timeCommitmentInSeconds - The length, in seconds, of the Shield Advanced subscription for the account.

$sel:subscriptionLimits:Subscription', subscription_subscriptionLimits - Limits settings for your subscription.

subscription_autoRenew :: Lens' Subscription (Maybe AutoRenew)

If ENABLED, the subscription will be automatically renewed at the end of the existing subscription period.

When you initally create a subscription, AutoRenew is set to ENABLED. You can change this by submitting an UpdateSubscription request. If the UpdateSubscription request does not included a value for AutoRenew, the existing value for AutoRenew remains unchanged.

subscription_proactiveEngagementStatus :: Lens' Subscription (Maybe ProactiveEngagementStatus)

If ENABLED, the Shield Response Team (SRT) will use email and phone to notify contacts about escalations to the SRT and to initiate proactive customer support.

If PENDING, you have requested proactive engagement and the request is pending. The status changes to ENABLED when your request is fully processed.

If DISABLED, the SRT will not proactively notify contacts about escalations or to initiate proactive customer support.

subscription_startTime :: Lens' Subscription (Maybe UTCTime)

The start time of the subscription, in Unix time in seconds. For more information see timestamp.

subscription_endTime :: Lens' Subscription (Maybe UTCTime)

The date and time your subscription will end.

subscription_limits :: Lens' Subscription (Maybe [Limit])

Specifies how many protections of a given type you can create.

subscription_subscriptionArn :: Lens' Subscription (Maybe Text)

The ARN (Amazon Resource Name) of the subscription.

subscription_timeCommitmentInSeconds :: Lens' Subscription (Maybe Natural)

The length, in seconds, of the Shield Advanced subscription for the account.

subscription_subscriptionLimits :: Lens' Subscription SubscriptionLimits

Limits settings for your subscription.

SubscriptionLimits

data SubscriptionLimits

Limits settings for your subscription.

See: newSubscriptionLimits smart constructor.

Constructors

SubscriptionLimits'

Fields

protectionLimits :: ProtectionLimits

Limits settings on protections for your subscription.

protectionGroupLimits :: ProtectionGroupLimits

Limits settings on protection groups for your subscription.

newSubscriptionLimits

Arguments

:: ProtectionLimits
$sel:protectionLimits:SubscriptionLimits'

-> ProtectionGroupLimits
$sel:protectionGroupLimits:SubscriptionLimits'

-> SubscriptionLimits

Create a value of SubscriptionLimits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protectionLimits:SubscriptionLimits', subscriptionLimits_protectionLimits - Limits settings on protections for your subscription.

$sel:protectionGroupLimits:SubscriptionLimits', subscriptionLimits_protectionGroupLimits - Limits settings on protection groups for your subscription.

subscriptionLimits_protectionLimits :: Lens' SubscriptionLimits ProtectionLimits

Limits settings on protections for your subscription.

subscriptionLimits_protectionGroupLimits :: Lens' SubscriptionLimits ProtectionGroupLimits

Limits settings on protection groups for your subscription.

SummarizedAttackVector

data SummarizedAttackVector

A summary of information about the attack.

See: newSummarizedAttackVector smart constructor.

Constructors

SummarizedAttackVector'

Fields

vectorCounters :: Maybe [SummarizedCounter]

The list of counters that describe the details of the attack.

vectorType :: Text

The attack type, for example, SNMP reflection or SYN flood.

newSummarizedAttackVector

Arguments

:: Text
$sel:vectorType:SummarizedAttackVector'

-> SummarizedAttackVector

Create a value of SummarizedAttackVector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vectorCounters:SummarizedAttackVector', summarizedAttackVector_vectorCounters - The list of counters that describe the details of the attack.

$sel:vectorType:SummarizedAttackVector', summarizedAttackVector_vectorType - The attack type, for example, SNMP reflection or SYN flood.

summarizedAttackVector_vectorCounters :: Lens' SummarizedAttackVector (Maybe [SummarizedCounter])

The list of counters that describe the details of the attack.

summarizedAttackVector_vectorType :: Lens' SummarizedAttackVector Text

The attack type, for example, SNMP reflection or SYN flood.

SummarizedCounter

data SummarizedCounter

The counter that describes a DDoS attack.

See: newSummarizedCounter smart constructor.

Constructors

SummarizedCounter'

Fields

n :: Maybe Int

The number of counters for a specified time period.

unit :: Maybe Text

The unit of the counters.

sum :: Maybe Double

The total of counter values for a specified time period.

name :: Maybe Text

The counter name.

max :: Maybe Double

The maximum value of the counter for a specified time period.

average :: Maybe Double

The average value of the counter for a specified time period.

newSummarizedCounter :: SummarizedCounter

Create a value of SummarizedCounter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:n:SummarizedCounter', summarizedCounter_n - The number of counters for a specified time period.

$sel:unit:SummarizedCounter', summarizedCounter_unit - The unit of the counters.

$sel:sum:SummarizedCounter', summarizedCounter_sum - The total of counter values for a specified time period.

$sel:name:SummarizedCounter', summarizedCounter_name - The counter name.

$sel:max:SummarizedCounter', summarizedCounter_max - The maximum value of the counter for a specified time period.

$sel:average:SummarizedCounter', summarizedCounter_average - The average value of the counter for a specified time period.

summarizedCounter_n :: Lens' SummarizedCounter (Maybe Int)

The number of counters for a specified time period.

summarizedCounter_unit :: Lens' SummarizedCounter (Maybe Text)

The unit of the counters.

summarizedCounter_sum :: Lens' SummarizedCounter (Maybe Double)

The total of counter values for a specified time period.

summarizedCounter_name :: Lens' SummarizedCounter (Maybe Text)

The counter name.

summarizedCounter_max :: Lens' SummarizedCounter (Maybe Double)

The maximum value of the counter for a specified time period.

summarizedCounter_average :: Lens' SummarizedCounter (Maybe Double)

The average value of the counter for a specified time period.

Tag

data Tag

A tag associated with an Amazon Web Services resource. Tags are key:value pairs that you can use to categorize and manage your resources, for purposes like billing or other management. Typically, the tag key represents a category, such as "environment", and the tag value represents a specific value within that category, such as "test," "development," or "production". Or you might set the tag key to "customer" and the value to the customer name or ID. You can specify one or more tags to add to each Amazon Web Services resource, up to 50 tags for a resource.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Maybe Text

Part of the key:value pair that defines a tag. You can use a tag key to describe a category of information, such as "customer." Tag keys are case-sensitive.

value :: Maybe Text

Part of the key:value pair that defines a tag. You can use a tag value to describe a specific value within a category, such as "companyA" or "companyB." Tag values are case-sensitive.

newTag :: Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - Part of the key:value pair that defines a tag. You can use a tag key to describe a category of information, such as "customer." Tag keys are case-sensitive.

$sel:value:Tag', tag_value - Part of the key:value pair that defines a tag. You can use a tag value to describe a specific value within a category, such as "companyA" or "companyB." Tag values are case-sensitive.

tag_key :: Lens' Tag (Maybe Text)

Part of the key:value pair that defines a tag. You can use a tag key to describe a category of information, such as "customer." Tag keys are case-sensitive.

tag_value :: Lens' Tag (Maybe Text)

Part of the key:value pair that defines a tag. You can use a tag value to describe a specific value within a category, such as "companyA" or "companyB." Tag values are case-sensitive.

TimeRange

data TimeRange

The time range.

See: newTimeRange smart constructor.

Constructors

TimeRange'

Fields

fromInclusive :: Maybe POSIX

The start time, in Unix time in seconds. For more information see timestamp.

toExclusive :: Maybe POSIX

The end time, in Unix time in seconds. For more information see timestamp.

newTimeRange :: TimeRange

Create a value of TimeRange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fromInclusive:TimeRange', timeRange_fromInclusive - The start time, in Unix time in seconds. For more information see timestamp.

$sel:toExclusive:TimeRange', timeRange_toExclusive - The end time, in Unix time in seconds. For more information see timestamp.

timeRange_fromInclusive :: Lens' TimeRange (Maybe UTCTime)

The start time, in Unix time in seconds. For more information see timestamp.

timeRange_toExclusive :: Lens' TimeRange (Maybe UTCTime)

The end time, in Unix time in seconds. For more information see timestamp.