Network.AWS.Shield.Types.AttackDetail

Description

Documentation

data AttackDetail

The details of a DDoS attack.

See: newAttackDetail smart constructor.

Constructors

AttackDetail'

Fields

attackId :: Maybe Text

The unique identifier (ID) of the attack.

startTime :: Maybe POSIX

The time the attack started, in Unix time in seconds. For more information see timestamp.

subResources :: Maybe [SubResourceSummary]

If applicable, additional detail about the resource being attacked, for example, IP address or URL.

mitigations :: Maybe [Mitigation]

List of mitigation actions taken for the attack.

attackProperties :: Maybe [AttackProperty]

The array of objects that provide details of the Shield event.

For infrastructure layer events (L3 and L4 events) after January 25, 2021, you can view metrics for top contributors in Amazon CloudWatch metrics. For more information, see Shield metrics and alarms in the WAF Developer Guide.

attackCounters :: Maybe [SummarizedCounter]

List of counters that describe the attack for the specified time period.

resourceArn :: Maybe Text

The ARN (Amazon Resource Name) of the resource that was attacked.

endTime :: Maybe POSIX

The time the attack ended, in Unix time in seconds. For more information see timestamp.

newAttackDetail :: AttackDetail

Create a value of AttackDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attackId:AttackDetail', attackDetail_attackId - The unique identifier (ID) of the attack.

$sel:startTime:AttackDetail', attackDetail_startTime - The time the attack started, in Unix time in seconds. For more information see timestamp.

$sel:subResources:AttackDetail', attackDetail_subResources - If applicable, additional detail about the resource being attacked, for example, IP address or URL.

$sel:mitigations:AttackDetail', attackDetail_mitigations - List of mitigation actions taken for the attack.

$sel:attackProperties:AttackDetail', attackDetail_attackProperties - The array of objects that provide details of the Shield event.

For infrastructure layer events (L3 and L4 events) after January 25, 2021, you can view metrics for top contributors in Amazon CloudWatch metrics. For more information, see Shield metrics and alarms in the WAF Developer Guide.

$sel:attackCounters:AttackDetail', attackDetail_attackCounters - List of counters that describe the attack for the specified time period.

$sel:resourceArn:AttackDetail', attackDetail_resourceArn - The ARN (Amazon Resource Name) of the resource that was attacked.

$sel:endTime:AttackDetail', attackDetail_endTime - The time the attack ended, in Unix time in seconds. For more information see timestamp.

attackDetail_attackId :: Lens' AttackDetail (Maybe Text)

The unique identifier (ID) of the attack.

attackDetail_startTime :: Lens' AttackDetail (Maybe UTCTime)

The time the attack started, in Unix time in seconds. For more information see timestamp.

attackDetail_subResources :: Lens' AttackDetail (Maybe [SubResourceSummary])

If applicable, additional detail about the resource being attacked, for example, IP address or URL.

attackDetail_mitigations :: Lens' AttackDetail (Maybe [Mitigation])

List of mitigation actions taken for the attack.

attackDetail_attackProperties :: Lens' AttackDetail (Maybe [AttackProperty])

The array of objects that provide details of the Shield event.

For infrastructure layer events (L3 and L4 events) after January 25, 2021, you can view metrics for top contributors in Amazon CloudWatch metrics. For more information, see Shield metrics and alarms in the WAF Developer Guide.

attackDetail_attackCounters :: Lens' AttackDetail (Maybe [SummarizedCounter])

List of counters that describe the attack for the specified time period.

attackDetail_resourceArn :: Lens' AttackDetail (Maybe Text)

The ARN (Amazon Resource Name) of the resource that was attacked.

attackDetail_endTime :: Lens' AttackDetail (Maybe UTCTime)

The time the attack ended, in Unix time in seconds. For more information see timestamp.