Network.AWS.Shield.Types.AttackStatisticsDataItem

Description

Documentation

data AttackStatisticsDataItem

A single attack statistics data record. This is returned by DescribeAttackStatistics along with a time range indicating the time period that the attack statistics apply to.

See: newAttackStatisticsDataItem smart constructor.

Constructors

AttackStatisticsDataItem'

Fields

attackVolume :: Maybe AttackVolume

Information about the volume of attacks during the time period. If the accompanying AttackCount is zero, this setting might be empty.

attackCount :: Integer

The number of attacks detected during the time period. This is always present, but might be zero.

newAttackStatisticsDataItem

Arguments

:: Integer
$sel:attackCount:AttackStatisticsDataItem'

-> AttackStatisticsDataItem

Create a value of AttackStatisticsDataItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attackVolume:AttackStatisticsDataItem', attackStatisticsDataItem_attackVolume - Information about the volume of attacks during the time period. If the accompanying AttackCount is zero, this setting might be empty.

$sel:attackCount:AttackStatisticsDataItem', attackStatisticsDataItem_attackCount - The number of attacks detected during the time period. This is always present, but might be zero.

attackStatisticsDataItem_attackVolume :: Lens' AttackStatisticsDataItem (Maybe AttackVolume)

Information about the volume of attacks during the time period. If the accompanying AttackCount is zero, this setting might be empty.

attackStatisticsDataItem_attackCount :: Lens' AttackStatisticsDataItem Integer

The number of attacks detected during the time period. This is always present, but might be zero.