Network.AWS.Shield.Types.AttackSummary

Description

Documentation

data AttackSummary

Summarizes all DDoS attacks for a specified time period.

See: newAttackSummary smart constructor.

Constructors

AttackSummary'

Fields

attackVectors :: Maybe [AttackVectorDescription]

The list of attacks for a specified time period.

attackId :: Maybe Text

The unique identifier (ID) of the attack.

startTime :: Maybe POSIX

The start time of the attack, in Unix time in seconds. For more information see timestamp.

resourceArn :: Maybe Text

The ARN (Amazon Resource Name) of the resource that was attacked.

endTime :: Maybe POSIX

The end time of the attack, in Unix time in seconds. For more information see timestamp.

newAttackSummary :: AttackSummary

Create a value of AttackSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attackVectors:AttackSummary', attackSummary_attackVectors - The list of attacks for a specified time period.

$sel:attackId:AttackSummary', attackSummary_attackId - The unique identifier (ID) of the attack.

$sel:startTime:AttackSummary', attackSummary_startTime - The start time of the attack, in Unix time in seconds. For more information see timestamp.

$sel:resourceArn:AttackSummary', attackSummary_resourceArn - The ARN (Amazon Resource Name) of the resource that was attacked.

$sel:endTime:AttackSummary', attackSummary_endTime - The end time of the attack, in Unix time in seconds. For more information see timestamp.

attackSummary_attackVectors :: Lens' AttackSummary (Maybe [AttackVectorDescription])

The list of attacks for a specified time period.

attackSummary_attackId :: Lens' AttackSummary (Maybe Text)

The unique identifier (ID) of the attack.

attackSummary_startTime :: Lens' AttackSummary (Maybe UTCTime)

The start time of the attack, in Unix time in seconds. For more information see timestamp.

attackSummary_resourceArn :: Lens' AttackSummary (Maybe Text)

The ARN (Amazon Resource Name) of the resource that was attacked.

attackSummary_endTime :: Lens' AttackSummary (Maybe UTCTime)

The end time of the attack, in Unix time in seconds. For more information see timestamp.