Network.AWS.Shield.Types.ProtectionGroup

Description

Documentation

data ProtectionGroup

A grouping of protected resources that you and Shield Advanced can monitor as a collective. This resource grouping improves the accuracy of detection and reduces false positives.

See: newProtectionGroup smart constructor.

Constructors

ProtectionGroup'

Fields

resourceType :: Maybe ProtectedResourceType

The resource type to include in the protection group. All protected resources of this type are included in the protection group. You must set this when you set Pattern to BY_RESOURCE_TYPE and you must not set it for any other Pattern setting.

protectionGroupArn :: Maybe Text

The ARN (Amazon Resource Name) of the protection group.

protectionGroupId :: Text

The name of the protection group. You use this to identify the protection group in lists and to manage the protection group, for example to update, delete, or describe it.

aggregation :: ProtectionGroupAggregation

Defines how Shield combines resource data for the group in order to detect, mitigate, and report events.

Sum - Use the total traffic across the group. This is a good choice for most cases. Examples include Elastic IP addresses for EC2 instances that scale manually or automatically.

Mean - Use the average of the traffic across the group. This is a good choice for resources that share traffic uniformly. Examples include accelerators and load balancers.

Max - Use the highest traffic from each resource. This is useful for resources that don't share traffic and for resources that share that traffic in a non-uniform way. Examples include Amazon CloudFront distributions and origin resources for CloudFront distributions.

pattern' :: ProtectionGroupPattern

The criteria to use to choose the protected resources for inclusion in the group. You can include all resources that have protections, provide a list of resource Amazon Resource Names (ARNs), or include all resources of a specified resource type.

members :: [Text]

The Amazon Resource Names (ARNs) of the resources to include in the protection group. You must set this when you set Pattern to ARBITRARY and you must not set it for any other Pattern setting.

newProtectionGroup

Arguments

:: Text
$sel:protectionGroupId:ProtectionGroup'

-> ProtectionGroupAggregation
$sel:aggregation:ProtectionGroup'

-> ProtectionGroupPattern
$sel:pattern':ProtectionGroup'

-> ProtectionGroup

Create a value of ProtectionGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ProtectionGroup', protectionGroup_resourceType - The resource type to include in the protection group. All protected resources of this type are included in the protection group. You must set this when you set Pattern to BY_RESOURCE_TYPE and you must not set it for any other Pattern setting.

$sel:protectionGroupArn:ProtectionGroup', protectionGroup_protectionGroupArn - The ARN (Amazon Resource Name) of the protection group.

$sel:protectionGroupId:ProtectionGroup', protectionGroup_protectionGroupId - The name of the protection group. You use this to identify the protection group in lists and to manage the protection group, for example to update, delete, or describe it.

$sel:aggregation:ProtectionGroup', protectionGroup_aggregation - Defines how Shield combines resource data for the group in order to detect, mitigate, and report events.

Sum - Use the total traffic across the group. This is a good choice for most cases. Examples include Elastic IP addresses for EC2 instances that scale manually or automatically.

Mean - Use the average of the traffic across the group. This is a good choice for resources that share traffic uniformly. Examples include accelerators and load balancers.

Max - Use the highest traffic from each resource. This is useful for resources that don't share traffic and for resources that share that traffic in a non-uniform way. Examples include Amazon CloudFront distributions and origin resources for CloudFront distributions.

$sel:pattern':ProtectionGroup', protectionGroup_pattern - The criteria to use to choose the protected resources for inclusion in the group. You can include all resources that have protections, provide a list of resource Amazon Resource Names (ARNs), or include all resources of a specified resource type.

$sel:members:ProtectionGroup', protectionGroup_members - The Amazon Resource Names (ARNs) of the resources to include in the protection group. You must set this when you set Pattern to ARBITRARY and you must not set it for any other Pattern setting.

protectionGroup_resourceType :: Lens' ProtectionGroup (Maybe ProtectedResourceType)

The resource type to include in the protection group. All protected resources of this type are included in the protection group. You must set this when you set Pattern to BY_RESOURCE_TYPE and you must not set it for any other Pattern setting.

protectionGroup_protectionGroupArn :: Lens' ProtectionGroup (Maybe Text)

The ARN (Amazon Resource Name) of the protection group.

protectionGroup_protectionGroupId :: Lens' ProtectionGroup Text

The name of the protection group. You use this to identify the protection group in lists and to manage the protection group, for example to update, delete, or describe it.

protectionGroup_aggregation :: Lens' ProtectionGroup ProtectionGroupAggregation

Defines how Shield combines resource data for the group in order to detect, mitigate, and report events.

Sum - Use the total traffic across the group. This is a good choice for most cases. Examples include Elastic IP addresses for EC2 instances that scale manually or automatically.

Mean - Use the average of the traffic across the group. This is a good choice for resources that share traffic uniformly. Examples include accelerators and load balancers.

Max - Use the highest traffic from each resource. This is useful for resources that don't share traffic and for resources that share that traffic in a non-uniform way. Examples include Amazon CloudFront distributions and origin resources for CloudFront distributions.

protectionGroup_pattern :: Lens' ProtectionGroup ProtectionGroupPattern

The criteria to use to choose the protected resources for inclusion in the group. You can include all resources that have protections, provide a list of resource Amazon Resource Names (ARNs), or include all resources of a specified resource type.

protectionGroup_members :: Lens' ProtectionGroup [Text]

The Amazon Resource Names (ARNs) of the resources to include in the protection group. You must set this when you set Pattern to ARBITRARY and you must not set it for any other Pattern setting.