Network.AWS.Shield.Types.Subscription

Description

Documentation

data Subscription

Information about the Shield Advanced subscription for an account.

See: newSubscription smart constructor.

Constructors

Subscription'

Fields

autoRenew :: Maybe AutoRenew

If ENABLED, the subscription will be automatically renewed at the end of the existing subscription period.

When you initally create a subscription, AutoRenew is set to ENABLED. You can change this by submitting an UpdateSubscription request. If the UpdateSubscription request does not included a value for AutoRenew, the existing value for AutoRenew remains unchanged.

proactiveEngagementStatus :: Maybe ProactiveEngagementStatus

If ENABLED, the Shield Response Team (SRT) will use email and phone to notify contacts about escalations to the SRT and to initiate proactive customer support.

If PENDING, you have requested proactive engagement and the request is pending. The status changes to ENABLED when your request is fully processed.

If DISABLED, the SRT will not proactively notify contacts about escalations or to initiate proactive customer support.

startTime :: Maybe POSIX

The start time of the subscription, in Unix time in seconds. For more information see timestamp.

endTime :: Maybe POSIX

The date and time your subscription will end.

limits :: Maybe [Limit]

Specifies how many protections of a given type you can create.

subscriptionArn :: Maybe Text

The ARN (Amazon Resource Name) of the subscription.

timeCommitmentInSeconds :: Maybe Natural

The length, in seconds, of the Shield Advanced subscription for the account.

subscriptionLimits :: SubscriptionLimits

Limits settings for your subscription.

newSubscription

Arguments

:: SubscriptionLimits
$sel:subscriptionLimits:Subscription'

-> Subscription

Create a value of Subscription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoRenew:Subscription', subscription_autoRenew - If ENABLED, the subscription will be automatically renewed at the end of the existing subscription period.

When you initally create a subscription, AutoRenew is set to ENABLED. You can change this by submitting an UpdateSubscription request. If the UpdateSubscription request does not included a value for AutoRenew, the existing value for AutoRenew remains unchanged.

$sel:proactiveEngagementStatus:Subscription', subscription_proactiveEngagementStatus - If ENABLED, the Shield Response Team (SRT) will use email and phone to notify contacts about escalations to the SRT and to initiate proactive customer support.

If PENDING, you have requested proactive engagement and the request is pending. The status changes to ENABLED when your request is fully processed.

If DISABLED, the SRT will not proactively notify contacts about escalations or to initiate proactive customer support.

$sel:startTime:Subscription', subscription_startTime - The start time of the subscription, in Unix time in seconds. For more information see timestamp.

$sel:endTime:Subscription', subscription_endTime - The date and time your subscription will end.

$sel:limits:Subscription', subscription_limits - Specifies how many protections of a given type you can create.

$sel:subscriptionArn:Subscription', subscription_subscriptionArn - The ARN (Amazon Resource Name) of the subscription.

$sel:timeCommitmentInSeconds:Subscription', subscription_timeCommitmentInSeconds - The length, in seconds, of the Shield Advanced subscription for the account.

$sel:subscriptionLimits:Subscription', subscription_subscriptionLimits - Limits settings for your subscription.

subscription_autoRenew :: Lens' Subscription (Maybe AutoRenew)

If ENABLED, the subscription will be automatically renewed at the end of the existing subscription period.

When you initally create a subscription, AutoRenew is set to ENABLED. You can change this by submitting an UpdateSubscription request. If the UpdateSubscription request does not included a value for AutoRenew, the existing value for AutoRenew remains unchanged.

subscription_proactiveEngagementStatus :: Lens' Subscription (Maybe ProactiveEngagementStatus)

If ENABLED, the Shield Response Team (SRT) will use email and phone to notify contacts about escalations to the SRT and to initiate proactive customer support.

If PENDING, you have requested proactive engagement and the request is pending. The status changes to ENABLED when your request is fully processed.

If DISABLED, the SRT will not proactively notify contacts about escalations or to initiate proactive customer support.

subscription_startTime :: Lens' Subscription (Maybe UTCTime)

The start time of the subscription, in Unix time in seconds. For more information see timestamp.

subscription_endTime :: Lens' Subscription (Maybe UTCTime)

The date and time your subscription will end.

subscription_limits :: Lens' Subscription (Maybe [Limit])

Specifies how many protections of a given type you can create.

subscription_subscriptionArn :: Lens' Subscription (Maybe Text)

The ARN (Amazon Resource Name) of the subscription.

subscription_timeCommitmentInSeconds :: Lens' Subscription (Maybe Natural)

The length, in seconds, of the Shield Advanced subscription for the account.

subscription_subscriptionLimits :: Lens' Subscription SubscriptionLimits

Limits settings for your subscription.