Network.AWS.Shield.Types.SummarizedCounter

Description

Documentation

data SummarizedCounter

The counter that describes a DDoS attack.

See: newSummarizedCounter smart constructor.

Constructors

SummarizedCounter'

Fields

n :: Maybe Int

The number of counters for a specified time period.

unit :: Maybe Text

The unit of the counters.

sum :: Maybe Double

The total of counter values for a specified time period.

name :: Maybe Text

The counter name.

max :: Maybe Double

The maximum value of the counter for a specified time period.

average :: Maybe Double

The average value of the counter for a specified time period.

newSummarizedCounter :: SummarizedCounter

Create a value of SummarizedCounter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:n:SummarizedCounter', summarizedCounter_n - The number of counters for a specified time period.

$sel:unit:SummarizedCounter', summarizedCounter_unit - The unit of the counters.

$sel:sum:SummarizedCounter', summarizedCounter_sum - The total of counter values for a specified time period.

$sel:name:SummarizedCounter', summarizedCounter_name - The counter name.

$sel:max:SummarizedCounter', summarizedCounter_max - The maximum value of the counter for a specified time period.

$sel:average:SummarizedCounter', summarizedCounter_average - The average value of the counter for a specified time period.

summarizedCounter_n :: Lens' SummarizedCounter (Maybe Int)

The number of counters for a specified time period.

summarizedCounter_unit :: Lens' SummarizedCounter (Maybe Text)

The unit of the counters.

summarizedCounter_sum :: Lens' SummarizedCounter (Maybe Double)

The total of counter values for a specified time period.

summarizedCounter_name :: Lens' SummarizedCounter (Maybe Text)

The counter name.

summarizedCounter_max :: Lens' SummarizedCounter (Maybe Double)

The maximum value of the counter for a specified time period.

summarizedCounter_average :: Lens' SummarizedCounter (Maybe Double)

The average value of the counter for a specified time period.