Network.AWS.Shield.UpdateProtectionGroup

Description

Updates an existing protection group. A protection group is a grouping of protected resources so they can be handled as a collective. This resource grouping improves the accuracy of detection and reduces false positives.

Creating a Request

data UpdateProtectionGroup

See: newUpdateProtectionGroup smart constructor.

Constructors

UpdateProtectionGroup'

Fields

resourceType :: Maybe ProtectedResourceType

The resource type to include in the protection group. All protected resources of this type are included in the protection group. You must set this when you set Pattern to BY_RESOURCE_TYPE and you must not set it for any other Pattern setting.

members :: Maybe [Text]

The Amazon Resource Names (ARNs) of the resources to include in the protection group. You must set this when you set Pattern to ARBITRARY and you must not set it for any other Pattern setting.

protectionGroupId :: Text

The name of the protection group. You use this to identify the protection group in lists and to manage the protection group, for example to update, delete, or describe it.

aggregation :: ProtectionGroupAggregation

Defines how Shield combines resource data for the group in order to detect, mitigate, and report events.

Sum - Use the total traffic across the group. This is a good choice for most cases. Examples include Elastic IP addresses for EC2 instances that scale manually or automatically.

Mean - Use the average of the traffic across the group. This is a good choice for resources that share traffic uniformly. Examples include accelerators and load balancers.

Max - Use the highest traffic from each resource. This is useful for resources that don't share traffic and for resources that share that traffic in a non-uniform way. Examples include Amazon CloudFront distributions and origin resources for CloudFront distributions.

pattern' :: ProtectionGroupPattern

The criteria to use to choose the protected resources for inclusion in the group. You can include all resources that have protections, provide a list of resource Amazon Resource Names (ARNs), or include all resources of a specified resource type.

newUpdateProtectionGroup

Arguments

:: Text
$sel:protectionGroupId:UpdateProtectionGroup'

-> ProtectionGroupAggregation
$sel:aggregation:UpdateProtectionGroup'

-> ProtectionGroupPattern
$sel:pattern':UpdateProtectionGroup'

-> UpdateProtectionGroup

Create a value of UpdateProtectionGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:UpdateProtectionGroup', updateProtectionGroup_resourceType - The resource type to include in the protection group. All protected resources of this type are included in the protection group. You must set this when you set Pattern to BY_RESOURCE_TYPE and you must not set it for any other Pattern setting.

$sel:members:UpdateProtectionGroup', updateProtectionGroup_members - The Amazon Resource Names (ARNs) of the resources to include in the protection group. You must set this when you set Pattern to ARBITRARY and you must not set it for any other Pattern setting.

$sel:protectionGroupId:UpdateProtectionGroup', updateProtectionGroup_protectionGroupId - The name of the protection group. You use this to identify the protection group in lists and to manage the protection group, for example to update, delete, or describe it.

$sel:aggregation:UpdateProtectionGroup', updateProtectionGroup_aggregation - Defines how Shield combines resource data for the group in order to detect, mitigate, and report events.

Sum - Use the total traffic across the group. This is a good choice for most cases. Examples include Elastic IP addresses for EC2 instances that scale manually or automatically.

Mean - Use the average of the traffic across the group. This is a good choice for resources that share traffic uniformly. Examples include accelerators and load balancers.

Max - Use the highest traffic from each resource. This is useful for resources that don't share traffic and for resources that share that traffic in a non-uniform way. Examples include Amazon CloudFront distributions and origin resources for CloudFront distributions.

$sel:pattern':UpdateProtectionGroup', updateProtectionGroup_pattern - The criteria to use to choose the protected resources for inclusion in the group. You can include all resources that have protections, provide a list of resource Amazon Resource Names (ARNs), or include all resources of a specified resource type.

Request Lenses

updateProtectionGroup_resourceType :: Lens' UpdateProtectionGroup (Maybe ProtectedResourceType)

The resource type to include in the protection group. All protected resources of this type are included in the protection group. You must set this when you set Pattern to BY_RESOURCE_TYPE and you must not set it for any other Pattern setting.

updateProtectionGroup_members :: Lens' UpdateProtectionGroup (Maybe [Text])

The Amazon Resource Names (ARNs) of the resources to include in the protection group. You must set this when you set Pattern to ARBITRARY and you must not set it for any other Pattern setting.

updateProtectionGroup_protectionGroupId :: Lens' UpdateProtectionGroup Text

The name of the protection group. You use this to identify the protection group in lists and to manage the protection group, for example to update, delete, or describe it.

updateProtectionGroup_aggregation :: Lens' UpdateProtectionGroup ProtectionGroupAggregation

Defines how Shield combines resource data for the group in order to detect, mitigate, and report events.

Sum - Use the total traffic across the group. This is a good choice for most cases. Examples include Elastic IP addresses for EC2 instances that scale manually or automatically.

Mean - Use the average of the traffic across the group. This is a good choice for resources that share traffic uniformly. Examples include accelerators and load balancers.

Max - Use the highest traffic from each resource. This is useful for resources that don't share traffic and for resources that share that traffic in a non-uniform way. Examples include Amazon CloudFront distributions and origin resources for CloudFront distributions.

updateProtectionGroup_pattern :: Lens' UpdateProtectionGroup ProtectionGroupPattern

The criteria to use to choose the protected resources for inclusion in the group. You can include all resources that have protections, provide a list of resource Amazon Resource Names (ARNs), or include all resources of a specified resource type.

Destructuring the Response

data UpdateProtectionGroupResponse

See: newUpdateProtectionGroupResponse smart constructor.

Constructors

UpdateProtectionGroupResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateProtectionGroupResponse

Arguments

:: Int
$sel:httpStatus:UpdateProtectionGroupResponse'

-> UpdateProtectionGroupResponse

Create a value of UpdateProtectionGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateProtectionGroupResponse', updateProtectionGroupResponse_httpStatus - The response's http status code.

Response Lenses

updateProtectionGroupResponse_httpStatus :: Lens' UpdateProtectionGroupResponse Int

The response's http status code.