Network.AWS.Shield.UpdateSubscription

Description

Updates the details of an existing subscription. Only enter values for parameters you want to change. Empty parameters are not updated.

Creating a Request

data UpdateSubscription

See: newUpdateSubscription smart constructor.

Constructors

UpdateSubscription'

Fields

autoRenew :: Maybe AutoRenew

When you initally create a subscription, AutoRenew is set to ENABLED. If ENABLED, the subscription will be automatically renewed at the end of the existing subscription period. You can change this by submitting an UpdateSubscription request. If the UpdateSubscription request does not included a value for AutoRenew, the existing value for AutoRenew remains unchanged.

newUpdateSubscription :: UpdateSubscription

Create a value of UpdateSubscription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoRenew:UpdateSubscription', updateSubscription_autoRenew - When you initally create a subscription, AutoRenew is set to ENABLED. If ENABLED, the subscription will be automatically renewed at the end of the existing subscription period. You can change this by submitting an UpdateSubscription request. If the UpdateSubscription request does not included a value for AutoRenew, the existing value for AutoRenew remains unchanged.

Request Lenses

updateSubscription_autoRenew :: Lens' UpdateSubscription (Maybe AutoRenew)

When you initally create a subscription, AutoRenew is set to ENABLED. If ENABLED, the subscription will be automatically renewed at the end of the existing subscription period. You can change this by submitting an UpdateSubscription request. If the UpdateSubscription request does not included a value for AutoRenew, the existing value for AutoRenew remains unchanged.

Destructuring the Response

data UpdateSubscriptionResponse

See: newUpdateSubscriptionResponse smart constructor.

Constructors

UpdateSubscriptionResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateSubscriptionResponse

Arguments

:: Int
$sel:httpStatus:UpdateSubscriptionResponse'

-> UpdateSubscriptionResponse

Create a value of UpdateSubscriptionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateSubscriptionResponse', updateSubscriptionResponse_httpStatus - The response's http status code.

Response Lenses

updateSubscriptionResponse_httpStatus :: Lens' UpdateSubscriptionResponse Int

The response's http status code.