Network.AWS.Sign.V2Header

Description

This module provides an AWS compliant V2 Header request signer. It is based heavily on boto, specifically boto's HmacAuthV1Handler AWS capable signer. AWS documentation is available here.

Notice: Limitations include an inability to sign with a security token and inability to overwrite the Date header with an expiry.

Documentation

v2Header :: Signer