Network.AWS.Signer.DescribeSigningJob

Description

Returns information about a specific code signing job. You specify the job by using the jobId value that is returned by the StartSigningJob operation.

Creating a Request

data DescribeSigningJob

See: newDescribeSigningJob smart constructor.

Constructors

DescribeSigningJob'

Fields

jobId :: Text

The ID of the signing job on input.

newDescribeSigningJob

Arguments

:: Text
$sel:jobId:DescribeSigningJob'

-> DescribeSigningJob

Create a value of DescribeSigningJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:DescribeSigningJob', describeSigningJob_jobId - The ID of the signing job on input.

Request Lenses

describeSigningJob_jobId :: Lens' DescribeSigningJob Text

The ID of the signing job on input.

Destructuring the Response

data DescribeSigningJobResponse

See: newDescribeSigningJobResponse smart constructor.

Constructors

DescribeSigningJobResponse'

Fields

status :: Maybe SigningStatus

Status of the signing job.

overrides :: Maybe SigningPlatformOverrides

A list of any overrides that were applied to the signing operation.

platformDisplayName :: Maybe Text

A human-readable name for the signing platform associated with the signing job.

jobId :: Maybe Text

The ID of the signing job on output.

createdAt :: Maybe POSIX

Date and time that the signing job was created.

signingMaterial :: Maybe SigningMaterial

The Amazon Resource Name (ARN) of your code signing certificate.

requestedBy :: Maybe Text

The IAM principal that requested the signing job.

signatureExpiresAt :: Maybe POSIX

Thr expiration timestamp for the signature generated by the signing job.

profileVersion :: Maybe Text

The version of the signing profile used to initiate the signing job.

profileName :: Maybe Text

The name of the profile that initiated the signing operation.

signedObject :: Maybe SignedObject

Name of the S3 bucket where the signed code image is saved by code signing.

platformId :: Maybe Text

The microcontroller platform to which your signed code image will be distributed.

source :: Maybe Source

The object that contains the name of your S3 bucket or your raw code.

revocationRecord :: Maybe SigningJobRevocationRecord

A revocation record if the signature generated by the signing job has been revoked. Contains a timestamp and the ID of the IAM entity that revoked the signature.

jobInvoker :: Maybe Text

The IAM entity that initiated the signing job.

completedAt :: Maybe POSIX

Date and time that the signing job was completed.

statusReason :: Maybe Text

String value that contains the status reason.

jobOwner :: Maybe Text

The AWS account ID of the job owner.

signingParameters :: Maybe (HashMap Text Text)

Map of user-assigned key-value pairs used during signing. These values contain any information that you specified for use in your signing job.

httpStatus :: Int

The response's http status code.

newDescribeSigningJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeSigningJobResponse'

-> DescribeSigningJobResponse

Create a value of DescribeSigningJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeSigningJobResponse', describeSigningJobResponse_status - Status of the signing job.

$sel:overrides:DescribeSigningJobResponse', describeSigningJobResponse_overrides - A list of any overrides that were applied to the signing operation.

$sel:platformDisplayName:DescribeSigningJobResponse', describeSigningJobResponse_platformDisplayName - A human-readable name for the signing platform associated with the signing job.

$sel:jobId:DescribeSigningJob', describeSigningJobResponse_jobId - The ID of the signing job on output.

$sel:createdAt:DescribeSigningJobResponse', describeSigningJobResponse_createdAt - Date and time that the signing job was created.

$sel:signingMaterial:DescribeSigningJobResponse', describeSigningJobResponse_signingMaterial - The Amazon Resource Name (ARN) of your code signing certificate.

$sel:requestedBy:DescribeSigningJobResponse', describeSigningJobResponse_requestedBy - The IAM principal that requested the signing job.

$sel:signatureExpiresAt:DescribeSigningJobResponse', describeSigningJobResponse_signatureExpiresAt - Thr expiration timestamp for the signature generated by the signing job.

$sel:profileVersion:DescribeSigningJobResponse', describeSigningJobResponse_profileVersion - The version of the signing profile used to initiate the signing job.

$sel:profileName:DescribeSigningJobResponse', describeSigningJobResponse_profileName - The name of the profile that initiated the signing operation.

$sel:signedObject:DescribeSigningJobResponse', describeSigningJobResponse_signedObject - Name of the S3 bucket where the signed code image is saved by code signing.

$sel:platformId:DescribeSigningJobResponse', describeSigningJobResponse_platformId - The microcontroller platform to which your signed code image will be distributed.

$sel:source:DescribeSigningJobResponse', describeSigningJobResponse_source - The object that contains the name of your S3 bucket or your raw code.

$sel:revocationRecord:DescribeSigningJobResponse', describeSigningJobResponse_revocationRecord - A revocation record if the signature generated by the signing job has been revoked. Contains a timestamp and the ID of the IAM entity that revoked the signature.

$sel:jobInvoker:DescribeSigningJobResponse', describeSigningJobResponse_jobInvoker - The IAM entity that initiated the signing job.

$sel:completedAt:DescribeSigningJobResponse', describeSigningJobResponse_completedAt - Date and time that the signing job was completed.

$sel:statusReason:DescribeSigningJobResponse', describeSigningJobResponse_statusReason - String value that contains the status reason.

$sel:jobOwner:DescribeSigningJobResponse', describeSigningJobResponse_jobOwner - The AWS account ID of the job owner.

$sel:signingParameters:DescribeSigningJobResponse', describeSigningJobResponse_signingParameters - Map of user-assigned key-value pairs used during signing. These values contain any information that you specified for use in your signing job.

$sel:httpStatus:DescribeSigningJobResponse', describeSigningJobResponse_httpStatus - The response's http status code.

Response Lenses

describeSigningJobResponse_status :: Lens' DescribeSigningJobResponse (Maybe SigningStatus)

Status of the signing job.

describeSigningJobResponse_overrides :: Lens' DescribeSigningJobResponse (Maybe SigningPlatformOverrides)

A list of any overrides that were applied to the signing operation.

describeSigningJobResponse_platformDisplayName :: Lens' DescribeSigningJobResponse (Maybe Text)

A human-readable name for the signing platform associated with the signing job.

describeSigningJobResponse_jobId :: Lens' DescribeSigningJobResponse (Maybe Text)

The ID of the signing job on output.

describeSigningJobResponse_createdAt :: Lens' DescribeSigningJobResponse (Maybe UTCTime)

Date and time that the signing job was created.

describeSigningJobResponse_signingMaterial :: Lens' DescribeSigningJobResponse (Maybe SigningMaterial)

The Amazon Resource Name (ARN) of your code signing certificate.

describeSigningJobResponse_requestedBy :: Lens' DescribeSigningJobResponse (Maybe Text)

The IAM principal that requested the signing job.

describeSigningJobResponse_signatureExpiresAt :: Lens' DescribeSigningJobResponse (Maybe UTCTime)

Thr expiration timestamp for the signature generated by the signing job.

describeSigningJobResponse_profileVersion :: Lens' DescribeSigningJobResponse (Maybe Text)

The version of the signing profile used to initiate the signing job.

describeSigningJobResponse_profileName :: Lens' DescribeSigningJobResponse (Maybe Text)

The name of the profile that initiated the signing operation.

describeSigningJobResponse_signedObject :: Lens' DescribeSigningJobResponse (Maybe SignedObject)

Name of the S3 bucket where the signed code image is saved by code signing.

describeSigningJobResponse_platformId :: Lens' DescribeSigningJobResponse (Maybe Text)

The microcontroller platform to which your signed code image will be distributed.

describeSigningJobResponse_source :: Lens' DescribeSigningJobResponse (Maybe Source)

The object that contains the name of your S3 bucket or your raw code.

describeSigningJobResponse_revocationRecord :: Lens' DescribeSigningJobResponse (Maybe SigningJobRevocationRecord)

A revocation record if the signature generated by the signing job has been revoked. Contains a timestamp and the ID of the IAM entity that revoked the signature.

describeSigningJobResponse_jobInvoker :: Lens' DescribeSigningJobResponse (Maybe Text)

The IAM entity that initiated the signing job.

describeSigningJobResponse_completedAt :: Lens' DescribeSigningJobResponse (Maybe UTCTime)

Date and time that the signing job was completed.

describeSigningJobResponse_statusReason :: Lens' DescribeSigningJobResponse (Maybe Text)

String value that contains the status reason.

describeSigningJobResponse_jobOwner :: Lens' DescribeSigningJobResponse (Maybe Text)

The AWS account ID of the job owner.

describeSigningJobResponse_signingParameters :: Lens' DescribeSigningJobResponse (Maybe (HashMap Text Text))

Map of user-assigned key-value pairs used during signing. These values contain any information that you specified for use in your signing job.

describeSigningJobResponse_httpStatus :: Lens' DescribeSigningJobResponse Int

The response's http status code.