Network.AWS.Signer.GetSigningPlatform

Description

Returns information on a specific signing platform.

Creating a Request

data GetSigningPlatform

See: newGetSigningPlatform smart constructor.

Constructors

GetSigningPlatform'

Fields

platformId :: Text

The ID of the target signing platform.

newGetSigningPlatform

Arguments

:: Text
$sel:platformId:GetSigningPlatform'

-> GetSigningPlatform

Create a value of GetSigningPlatform with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platformId:GetSigningPlatform', getSigningPlatform_platformId - The ID of the target signing platform.

Request Lenses

getSigningPlatform_platformId :: Lens' GetSigningPlatform Text

The ID of the target signing platform.

Destructuring the Response

data GetSigningPlatformResponse

See: newGetSigningPlatformResponse smart constructor.

Constructors

GetSigningPlatformResponse'

Fields

category :: Maybe Category

The category type of the target signing platform.

signingConfiguration :: Maybe SigningConfiguration

A list of configurations applied to the target platform at signing.

partner :: Maybe Text

A list of partner entities that use the target signing platform.

revocationSupported :: Maybe Bool

A flag indicating whether signatures generated for the signing platform can be revoked.

signingImageFormat :: Maybe SigningImageFormat

The format of the target platform's signing image.

platformId :: Maybe Text

The ID of the target signing platform.

displayName :: Maybe Text

The display name of the target signing platform.

maxSizeInMB :: Maybe Int

The maximum size (in MB) of the payload that can be signed by the target platform.

target :: Maybe Text

The validation template that is used by the target signing platform.

httpStatus :: Int

The response's http status code.

newGetSigningPlatformResponse

Arguments

:: Int
$sel:httpStatus:GetSigningPlatformResponse'

-> GetSigningPlatformResponse

Create a value of GetSigningPlatformResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:category:GetSigningPlatformResponse', getSigningPlatformResponse_category - The category type of the target signing platform.

$sel:signingConfiguration:GetSigningPlatformResponse', getSigningPlatformResponse_signingConfiguration - A list of configurations applied to the target platform at signing.

$sel:partner:GetSigningPlatformResponse', getSigningPlatformResponse_partner - A list of partner entities that use the target signing platform.

$sel:revocationSupported:GetSigningPlatformResponse', getSigningPlatformResponse_revocationSupported - A flag indicating whether signatures generated for the signing platform can be revoked.

$sel:signingImageFormat:GetSigningPlatformResponse', getSigningPlatformResponse_signingImageFormat - The format of the target platform's signing image.

$sel:platformId:GetSigningPlatform', getSigningPlatformResponse_platformId - The ID of the target signing platform.

$sel:displayName:GetSigningPlatformResponse', getSigningPlatformResponse_displayName - The display name of the target signing platform.

$sel:maxSizeInMB:GetSigningPlatformResponse', getSigningPlatformResponse_maxSizeInMB - The maximum size (in MB) of the payload that can be signed by the target platform.

$sel:target:GetSigningPlatformResponse', getSigningPlatformResponse_target - The validation template that is used by the target signing platform.

$sel:httpStatus:GetSigningPlatformResponse', getSigningPlatformResponse_httpStatus - The response's http status code.

Response Lenses

getSigningPlatformResponse_category :: Lens' GetSigningPlatformResponse (Maybe Category)

The category type of the target signing platform.

getSigningPlatformResponse_signingConfiguration :: Lens' GetSigningPlatformResponse (Maybe SigningConfiguration)

A list of configurations applied to the target platform at signing.

getSigningPlatformResponse_partner :: Lens' GetSigningPlatformResponse (Maybe Text)

A list of partner entities that use the target signing platform.

getSigningPlatformResponse_revocationSupported :: Lens' GetSigningPlatformResponse (Maybe Bool)

A flag indicating whether signatures generated for the signing platform can be revoked.

getSigningPlatformResponse_signingImageFormat :: Lens' GetSigningPlatformResponse (Maybe SigningImageFormat)

The format of the target platform's signing image.

getSigningPlatformResponse_platformId :: Lens' GetSigningPlatformResponse (Maybe Text)

The ID of the target signing platform.

getSigningPlatformResponse_displayName :: Lens' GetSigningPlatformResponse (Maybe Text)

The display name of the target signing platform.

getSigningPlatformResponse_maxSizeInMB :: Lens' GetSigningPlatformResponse (Maybe Int)

The maximum size (in MB) of the payload that can be signed by the target platform.

getSigningPlatformResponse_target :: Lens' GetSigningPlatformResponse (Maybe Text)

The validation template that is used by the target signing platform.

getSigningPlatformResponse_httpStatus :: Lens' GetSigningPlatformResponse Int

The response's http status code.