Network.AWS.Signer.GetSigningProfile

Description

Returns information on a specific signing profile.

Creating a Request

data GetSigningProfile

See: newGetSigningProfile smart constructor.

Constructors

GetSigningProfile'

Fields

profileOwner :: Maybe Text

The AWS account ID of the profile owner.

profileName :: Text

The name of the target signing profile.

newGetSigningProfile

Arguments

:: Text
$sel:profileName:GetSigningProfile'

-> GetSigningProfile

Create a value of GetSigningProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:profileOwner:GetSigningProfile', getSigningProfile_profileOwner - The AWS account ID of the profile owner.

$sel:profileName:GetSigningProfile', getSigningProfile_profileName - The name of the target signing profile.

Request Lenses

getSigningProfile_profileOwner :: Lens' GetSigningProfile (Maybe Text)

The AWS account ID of the profile owner.

getSigningProfile_profileName :: Lens' GetSigningProfile Text

The name of the target signing profile.

Destructuring the Response

data GetSigningProfileResponse

See: newGetSigningProfileResponse smart constructor.

Constructors

GetSigningProfileResponse'

Fields

status :: Maybe SigningProfileStatus

The status of the target signing profile.

overrides :: Maybe SigningPlatformOverrides

A list of overrides applied by the target signing profile for signing operations.

platformDisplayName :: Maybe Text

A human-readable name for the signing platform associated with the signing profile.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the signing profile.

signingMaterial :: Maybe SigningMaterial

The ARN of the certificate that the target profile uses for signing operations.

profileVersion :: Maybe Text

The current version of the signing profile.

profileName :: Maybe Text

The name of the target signing profile.

profileVersionArn :: Maybe Text

The signing profile ARN, including the profile version.

platformId :: Maybe Text

The ID of the platform that is used by the target signing profile.

revocationRecord :: Maybe SigningProfileRevocationRecord

statusReason :: Maybe Text

Reason for the status of the target signing profile.

signatureValidityPeriod :: Maybe SignatureValidityPeriod

signingParameters :: Maybe (HashMap Text Text)

A map of key-value pairs for signing operations that is attached to the target signing profile.

tags :: Maybe (HashMap Text Text)

A list of tags associated with the signing profile.

httpStatus :: Int

The response's http status code.

newGetSigningProfileResponse

Arguments

:: Int
$sel:httpStatus:GetSigningProfileResponse'

-> GetSigningProfileResponse

Create a value of GetSigningProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetSigningProfileResponse', getSigningProfileResponse_status - The status of the target signing profile.

$sel:overrides:GetSigningProfileResponse', getSigningProfileResponse_overrides - A list of overrides applied by the target signing profile for signing operations.

$sel:platformDisplayName:GetSigningProfileResponse', getSigningProfileResponse_platformDisplayName - A human-readable name for the signing platform associated with the signing profile.

$sel:arn:GetSigningProfileResponse', getSigningProfileResponse_arn - The Amazon Resource Name (ARN) for the signing profile.

$sel:signingMaterial:GetSigningProfileResponse', getSigningProfileResponse_signingMaterial - The ARN of the certificate that the target profile uses for signing operations.

$sel:profileVersion:GetSigningProfileResponse', getSigningProfileResponse_profileVersion - The current version of the signing profile.

$sel:profileName:GetSigningProfile', getSigningProfileResponse_profileName - The name of the target signing profile.

$sel:profileVersionArn:GetSigningProfileResponse', getSigningProfileResponse_profileVersionArn - The signing profile ARN, including the profile version.

$sel:platformId:GetSigningProfileResponse', getSigningProfileResponse_platformId - The ID of the platform that is used by the target signing profile.

$sel:revocationRecord:GetSigningProfileResponse', getSigningProfileResponse_revocationRecord - Undocumented member.

$sel:statusReason:GetSigningProfileResponse', getSigningProfileResponse_statusReason - Reason for the status of the target signing profile.

$sel:signatureValidityPeriod:GetSigningProfileResponse', getSigningProfileResponse_signatureValidityPeriod - Undocumented member.

$sel:signingParameters:GetSigningProfileResponse', getSigningProfileResponse_signingParameters - A map of key-value pairs for signing operations that is attached to the target signing profile.

$sel:tags:GetSigningProfileResponse', getSigningProfileResponse_tags - A list of tags associated with the signing profile.

$sel:httpStatus:GetSigningProfileResponse', getSigningProfileResponse_httpStatus - The response's http status code.

Response Lenses

getSigningProfileResponse_status :: Lens' GetSigningProfileResponse (Maybe SigningProfileStatus)

The status of the target signing profile.

getSigningProfileResponse_overrides :: Lens' GetSigningProfileResponse (Maybe SigningPlatformOverrides)

A list of overrides applied by the target signing profile for signing operations.

getSigningProfileResponse_platformDisplayName :: Lens' GetSigningProfileResponse (Maybe Text)

A human-readable name for the signing platform associated with the signing profile.

getSigningProfileResponse_arn :: Lens' GetSigningProfileResponse (Maybe Text)

The Amazon Resource Name (ARN) for the signing profile.

getSigningProfileResponse_signingMaterial :: Lens' GetSigningProfileResponse (Maybe SigningMaterial)

The ARN of the certificate that the target profile uses for signing operations.

getSigningProfileResponse_profileVersion :: Lens' GetSigningProfileResponse (Maybe Text)

The current version of the signing profile.

getSigningProfileResponse_profileName :: Lens' GetSigningProfileResponse (Maybe Text)

The name of the target signing profile.

getSigningProfileResponse_profileVersionArn :: Lens' GetSigningProfileResponse (Maybe Text)

The signing profile ARN, including the profile version.

getSigningProfileResponse_platformId :: Lens' GetSigningProfileResponse (Maybe Text)

The ID of the platform that is used by the target signing profile.

getSigningProfileResponse_revocationRecord :: Lens' GetSigningProfileResponse (Maybe SigningProfileRevocationRecord)

Undocumented member.

getSigningProfileResponse_statusReason :: Lens' GetSigningProfileResponse (Maybe Text)

Reason for the status of the target signing profile.

getSigningProfileResponse_signatureValidityPeriod :: Lens' GetSigningProfileResponse (Maybe SignatureValidityPeriod)

Undocumented member.

getSigningProfileResponse_signingParameters :: Lens' GetSigningProfileResponse (Maybe (HashMap Text Text))

A map of key-value pairs for signing operations that is attached to the target signing profile.

getSigningProfileResponse_tags :: Lens' GetSigningProfileResponse (Maybe (HashMap Text Text))

A list of tags associated with the signing profile.

getSigningProfileResponse_httpStatus :: Lens' GetSigningProfileResponse Int

The response's http status code.