Network.AWS.Signer.Lens

Description

Operations

StartSigningJob

startSigningJob_profileOwner :: Lens' StartSigningJob (Maybe Text)

The AWS account ID of the signing profile owner.

startSigningJob_source :: Lens' StartSigningJob Source

The S3 bucket that contains the object to sign or a BLOB that contains your raw code.

startSigningJob_destination :: Lens' StartSigningJob Destination

The S3 bucket in which to save your signed object. The destination contains the name of your bucket and an optional prefix.

startSigningJob_profileName :: Lens' StartSigningJob Text

The name of the signing profile.

startSigningJob_clientRequestToken :: Lens' StartSigningJob Text

String that identifies the signing request. All calls after the first that use this token return the same response as the first call.

startSigningJobResponse_jobId :: Lens' StartSigningJobResponse (Maybe Text)

The ID of your signing job.

startSigningJobResponse_jobOwner :: Lens' StartSigningJobResponse (Maybe Text)

The AWS account ID of the signing job owner.

startSigningJobResponse_httpStatus :: Lens' StartSigningJobResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) for the signing profile.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

A list of tags associated with the signing profile.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

RevokeSigningProfile

revokeSigningProfile_profileVersion :: Lens' RevokeSigningProfile Text

The version of the signing profile to be revoked.

revokeSigningProfile_reason :: Lens' RevokeSigningProfile Text

The reason for revoking a signing profile.

revokeSigningProfile_effectiveTime :: Lens' RevokeSigningProfile UTCTime

A timestamp for when revocation of a Signing Profile should become effective. Signatures generated using the signing profile after this timestamp are not trusted.

revokeSigningProfile_profileName :: Lens' RevokeSigningProfile Text

The name of the signing profile to be revoked.

CancelSigningProfile

cancelSigningProfile_profileName :: Lens' CancelSigningProfile Text

The name of the signing profile to be canceled.

PutSigningProfile

putSigningProfile_overrides :: Lens' PutSigningProfile (Maybe SigningPlatformOverrides)

A subfield of platform. This specifies any different configuration options that you want to apply to the chosen platform (such as a different hash-algorithm or signing-algorithm).

putSigningProfile_signingMaterial :: Lens' PutSigningProfile (Maybe SigningMaterial)

The AWS Certificate Manager certificate that will be used to sign code with the new signing profile.

putSigningProfile_signatureValidityPeriod :: Lens' PutSigningProfile (Maybe SignatureValidityPeriod)

The default validity period override for any signature generated using this signing profile. If unspecified, the default is 135 months.

putSigningProfile_signingParameters :: Lens' PutSigningProfile (Maybe (HashMap Text Text))

Map of key-value pairs for signing. These can include any information that you want to use during signing.

putSigningProfile_tags :: Lens' PutSigningProfile (Maybe (HashMap Text Text))

Tags to be associated with the signing profile that is being created.

putSigningProfile_profileName :: Lens' PutSigningProfile Text

The name of the signing profile to be created.

putSigningProfile_platformId :: Lens' PutSigningProfile Text

The ID of the signing platform to be created.

putSigningProfileResponse_arn :: Lens' PutSigningProfileResponse (Maybe Text)

The Amazon Resource Name (ARN) of the signing profile created.

putSigningProfileResponse_profileVersion :: Lens' PutSigningProfileResponse (Maybe Text)

The version of the signing profile being created.

putSigningProfileResponse_profileVersionArn :: Lens' PutSigningProfileResponse (Maybe Text)

The signing profile ARN, including the profile version.

putSigningProfileResponse_httpStatus :: Lens' PutSigningProfileResponse Int

The response's http status code.

AddProfilePermission

addProfilePermission_profileVersion :: Lens' AddProfilePermission (Maybe Text)

The version of the signing profile.

addProfilePermission_revisionId :: Lens' AddProfilePermission (Maybe Text)

A unique identifier for the current profile revision.

addProfilePermission_action :: Lens' AddProfilePermission Text

The AWS Signer action permitted as part of cross-account permissions.

addProfilePermission_principal :: Lens' AddProfilePermission Text

The AWS principal receiving cross-account permissions. This may be an IAM role or another AWS account ID.

addProfilePermission_statementId :: Lens' AddProfilePermission Text

A unique identifier for the cross-account permission statement.

addProfilePermission_profileName :: Lens' AddProfilePermission Text

The human-readable name of the signing profile.

addProfilePermissionResponse_revisionId :: Lens' AddProfilePermissionResponse (Maybe Text)

A unique identifier for the current profile revision.

addProfilePermissionResponse_httpStatus :: Lens' AddProfilePermissionResponse Int

The response's http status code.

ListSigningProfiles

listSigningProfiles_nextToken :: Lens' ListSigningProfiles (Maybe Text)

Value for specifying the next set of paginated results to return. After you receive a response with truncated results, use this parameter in a subsequent request. Set it to the value of nextToken from the response that you just received.

listSigningProfiles_platformId :: Lens' ListSigningProfiles (Maybe Text)

Filters results to return only signing jobs initiated for a specified signing platform.

listSigningProfiles_statuses :: Lens' ListSigningProfiles (Maybe [SigningProfileStatus])

Filters results to return only signing jobs with statuses in the specified list.

listSigningProfiles_includeCanceled :: Lens' ListSigningProfiles (Maybe Bool)

Designates whether to include profiles with the status of CANCELED.

listSigningProfiles_maxResults :: Lens' ListSigningProfiles (Maybe Natural)

The maximum number of profiles to be returned.

listSigningProfilesResponse_profiles :: Lens' ListSigningProfilesResponse (Maybe [SigningProfile])

A list of profiles that are available in the AWS account. This includes profiles with the status of CANCELED if the includeCanceled parameter is set to true.

listSigningProfilesResponse_nextToken :: Lens' ListSigningProfilesResponse (Maybe Text)

Value for specifying the next set of paginated results to return.

listSigningProfilesResponse_httpStatus :: Lens' ListSigningProfilesResponse Int

The response's http status code.

ListProfilePermissions

listProfilePermissions_nextToken :: Lens' ListProfilePermissions (Maybe Text)

String for specifying the next set of paginated results.

listProfilePermissions_profileName :: Lens' ListProfilePermissions Text

Name of the signing profile containing the cross-account permissions.

listProfilePermissionsResponse_policySizeBytes :: Lens' ListProfilePermissionsResponse (Maybe Int)

Total size of the policy associated with the Signing Profile in bytes.

listProfilePermissionsResponse_nextToken :: Lens' ListProfilePermissionsResponse (Maybe Text)

String for specifying the next set of paginated results.

listProfilePermissionsResponse_permissions :: Lens' ListProfilePermissionsResponse (Maybe [Permission])

List of permissions associated with the Signing Profile.

listProfilePermissionsResponse_revisionId :: Lens' ListProfilePermissionsResponse (Maybe Text)

The identifier for the current revision of profile permissions.

listProfilePermissionsResponse_httpStatus :: Lens' ListProfilePermissionsResponse Int

The response's http status code.

RevokeSignature

revokeSignature_jobOwner :: Lens' RevokeSignature (Maybe Text)

AWS account ID of the job owner.

revokeSignature_reason :: Lens' RevokeSignature Text

The reason for revoking the signing job.

revokeSignature_jobId :: Lens' RevokeSignature Text

ID of the signing job to be revoked.

GetSigningPlatform

getSigningPlatform_platformId :: Lens' GetSigningPlatform Text

The ID of the target signing platform.

getSigningPlatformResponse_category :: Lens' GetSigningPlatformResponse (Maybe Category)

The category type of the target signing platform.

getSigningPlatformResponse_signingConfiguration :: Lens' GetSigningPlatformResponse (Maybe SigningConfiguration)

A list of configurations applied to the target platform at signing.

getSigningPlatformResponse_partner :: Lens' GetSigningPlatformResponse (Maybe Text)

A list of partner entities that use the target signing platform.

getSigningPlatformResponse_revocationSupported :: Lens' GetSigningPlatformResponse (Maybe Bool)

A flag indicating whether signatures generated for the signing platform can be revoked.

getSigningPlatformResponse_signingImageFormat :: Lens' GetSigningPlatformResponse (Maybe SigningImageFormat)

The format of the target platform's signing image.

getSigningPlatformResponse_platformId :: Lens' GetSigningPlatformResponse (Maybe Text)

The ID of the target signing platform.

getSigningPlatformResponse_displayName :: Lens' GetSigningPlatformResponse (Maybe Text)

The display name of the target signing platform.

getSigningPlatformResponse_maxSizeInMB :: Lens' GetSigningPlatformResponse (Maybe Int)

The maximum size (in MB) of the payload that can be signed by the target platform.

getSigningPlatformResponse_target :: Lens' GetSigningPlatformResponse (Maybe Text)

The validation template that is used by the target signing platform.

getSigningPlatformResponse_httpStatus :: Lens' GetSigningPlatformResponse Int

The response's http status code.

ListSigningPlatforms

listSigningPlatforms_category :: Lens' ListSigningPlatforms (Maybe Text)

The category type of a signing platform.

listSigningPlatforms_partner :: Lens' ListSigningPlatforms (Maybe Text)

Any partner entities connected to a signing platform.

listSigningPlatforms_nextToken :: Lens' ListSigningPlatforms (Maybe Text)

Value for specifying the next set of paginated results to return. After you receive a response with truncated results, use this parameter in a subsequent request. Set it to the value of nextToken from the response that you just received.

listSigningPlatforms_maxResults :: Lens' ListSigningPlatforms (Maybe Natural)

The maximum number of results to be returned by this operation.

listSigningPlatforms_target :: Lens' ListSigningPlatforms (Maybe Text)

The validation template that is used by the target signing platform.

listSigningPlatformsResponse_platforms :: Lens' ListSigningPlatformsResponse (Maybe [SigningPlatform])

A list of all platforms that match the request parameters.

listSigningPlatformsResponse_nextToken :: Lens' ListSigningPlatformsResponse (Maybe Text)

Value for specifying the next set of paginated results to return.

listSigningPlatformsResponse_httpStatus :: Lens' ListSigningPlatformsResponse Int

The response's http status code.

ListSigningJobs

listSigningJobs_status :: Lens' ListSigningJobs (Maybe SigningStatus)

A status value with which to filter your results.

listSigningJobs_signatureExpiresAfter :: Lens' ListSigningJobs (Maybe UTCTime)

Filters results to return only signing jobs with signatures expiring after a specified timestamp.

listSigningJobs_requestedBy :: Lens' ListSigningJobs (Maybe Text)

The IAM principal that requested the signing job.

listSigningJobs_isRevoked :: Lens' ListSigningJobs (Maybe Bool)

Filters results to return only signing jobs with revoked signatures.

listSigningJobs_nextToken :: Lens' ListSigningJobs (Maybe Text)

String for specifying the next set of paginated results to return. After you receive a response with truncated results, use this parameter in a subsequent request. Set it to the value of nextToken from the response that you just received.

listSigningJobs_platformId :: Lens' ListSigningJobs (Maybe Text)

The ID of microcontroller platform that you specified for the distribution of your code image.

listSigningJobs_jobInvoker :: Lens' ListSigningJobs (Maybe Text)

Filters results to return only signing jobs initiated by a specified IAM entity.

listSigningJobs_signatureExpiresBefore :: Lens' ListSigningJobs (Maybe UTCTime)

Filters results to return only signing jobs with signatures expiring before a specified timestamp.

listSigningJobs_maxResults :: Lens' ListSigningJobs (Maybe Natural)

Specifies the maximum number of items to return in the response. Use this parameter when paginating results. If additional items exist beyond the number you specify, the nextToken element is set in the response. Use the nextToken value in a subsequent request to retrieve additional items.

listSigningJobsResponse_jobs :: Lens' ListSigningJobsResponse (Maybe [SigningJob])

A list of your signing jobs.

listSigningJobsResponse_nextToken :: Lens' ListSigningJobsResponse (Maybe Text)

String for specifying the next set of paginated results.

listSigningJobsResponse_httpStatus :: Lens' ListSigningJobsResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) for the signing profile.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

One or more tags to be associated with the signing profile.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) for the signing profile.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

A list of tag keys to be removed from the signing profile.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

RemoveProfilePermission

removeProfilePermission_revisionId :: Lens' RemoveProfilePermission Text

An identifier for the current revision of the signing profile permissions.

removeProfilePermission_profileName :: Lens' RemoveProfilePermission Text

A human-readable name for the signing profile with permissions to be removed.

removeProfilePermission_statementId :: Lens' RemoveProfilePermission Text

A unique identifier for the cross-account permissions statement.

removeProfilePermissionResponse_revisionId :: Lens' RemoveProfilePermissionResponse (Maybe Text)

An identifier for the current revision of the profile permissions.

removeProfilePermissionResponse_httpStatus :: Lens' RemoveProfilePermissionResponse Int

The response's http status code.

GetSigningProfile

getSigningProfile_profileOwner :: Lens' GetSigningProfile (Maybe Text)

The AWS account ID of the profile owner.

getSigningProfile_profileName :: Lens' GetSigningProfile Text

The name of the target signing profile.

getSigningProfileResponse_status :: Lens' GetSigningProfileResponse (Maybe SigningProfileStatus)

The status of the target signing profile.

getSigningProfileResponse_overrides :: Lens' GetSigningProfileResponse (Maybe SigningPlatformOverrides)

A list of overrides applied by the target signing profile for signing operations.

getSigningProfileResponse_platformDisplayName :: Lens' GetSigningProfileResponse (Maybe Text)

A human-readable name for the signing platform associated with the signing profile.

getSigningProfileResponse_arn :: Lens' GetSigningProfileResponse (Maybe Text)

The Amazon Resource Name (ARN) for the signing profile.

getSigningProfileResponse_signingMaterial :: Lens' GetSigningProfileResponse (Maybe SigningMaterial)

The ARN of the certificate that the target profile uses for signing operations.

getSigningProfileResponse_profileVersion :: Lens' GetSigningProfileResponse (Maybe Text)

The current version of the signing profile.

getSigningProfileResponse_profileName :: Lens' GetSigningProfileResponse (Maybe Text)

The name of the target signing profile.

getSigningProfileResponse_profileVersionArn :: Lens' GetSigningProfileResponse (Maybe Text)

The signing profile ARN, including the profile version.

getSigningProfileResponse_platformId :: Lens' GetSigningProfileResponse (Maybe Text)

The ID of the platform that is used by the target signing profile.

getSigningProfileResponse_revocationRecord :: Lens' GetSigningProfileResponse (Maybe SigningProfileRevocationRecord)

Undocumented member.

getSigningProfileResponse_statusReason :: Lens' GetSigningProfileResponse (Maybe Text)

Reason for the status of the target signing profile.

getSigningProfileResponse_signatureValidityPeriod :: Lens' GetSigningProfileResponse (Maybe SignatureValidityPeriod)

Undocumented member.

getSigningProfileResponse_signingParameters :: Lens' GetSigningProfileResponse (Maybe (HashMap Text Text))

A map of key-value pairs for signing operations that is attached to the target signing profile.

getSigningProfileResponse_tags :: Lens' GetSigningProfileResponse (Maybe (HashMap Text Text))

A list of tags associated with the signing profile.

getSigningProfileResponse_httpStatus :: Lens' GetSigningProfileResponse Int

The response's http status code.

DescribeSigningJob

describeSigningJob_jobId :: Lens' DescribeSigningJob Text

The ID of the signing job on input.

describeSigningJobResponse_status :: Lens' DescribeSigningJobResponse (Maybe SigningStatus)

Status of the signing job.

describeSigningJobResponse_overrides :: Lens' DescribeSigningJobResponse (Maybe SigningPlatformOverrides)

A list of any overrides that were applied to the signing operation.

describeSigningJobResponse_platformDisplayName :: Lens' DescribeSigningJobResponse (Maybe Text)

A human-readable name for the signing platform associated with the signing job.

describeSigningJobResponse_jobId :: Lens' DescribeSigningJobResponse (Maybe Text)

The ID of the signing job on output.

describeSigningJobResponse_createdAt :: Lens' DescribeSigningJobResponse (Maybe UTCTime)

Date and time that the signing job was created.

describeSigningJobResponse_signingMaterial :: Lens' DescribeSigningJobResponse (Maybe SigningMaterial)

The Amazon Resource Name (ARN) of your code signing certificate.

describeSigningJobResponse_requestedBy :: Lens' DescribeSigningJobResponse (Maybe Text)

The IAM principal that requested the signing job.

describeSigningJobResponse_signatureExpiresAt :: Lens' DescribeSigningJobResponse (Maybe UTCTime)

Thr expiration timestamp for the signature generated by the signing job.

describeSigningJobResponse_profileVersion :: Lens' DescribeSigningJobResponse (Maybe Text)

The version of the signing profile used to initiate the signing job.

describeSigningJobResponse_profileName :: Lens' DescribeSigningJobResponse (Maybe Text)

The name of the profile that initiated the signing operation.

describeSigningJobResponse_signedObject :: Lens' DescribeSigningJobResponse (Maybe SignedObject)

Name of the S3 bucket where the signed code image is saved by code signing.

describeSigningJobResponse_platformId :: Lens' DescribeSigningJobResponse (Maybe Text)

The microcontroller platform to which your signed code image will be distributed.

describeSigningJobResponse_source :: Lens' DescribeSigningJobResponse (Maybe Source)

The object that contains the name of your S3 bucket or your raw code.

describeSigningJobResponse_revocationRecord :: Lens' DescribeSigningJobResponse (Maybe SigningJobRevocationRecord)

A revocation record if the signature generated by the signing job has been revoked. Contains a timestamp and the ID of the IAM entity that revoked the signature.

describeSigningJobResponse_jobInvoker :: Lens' DescribeSigningJobResponse (Maybe Text)

The IAM entity that initiated the signing job.

describeSigningJobResponse_completedAt :: Lens' DescribeSigningJobResponse (Maybe UTCTime)

Date and time that the signing job was completed.

describeSigningJobResponse_statusReason :: Lens' DescribeSigningJobResponse (Maybe Text)

String value that contains the status reason.

describeSigningJobResponse_jobOwner :: Lens' DescribeSigningJobResponse (Maybe Text)

The AWS account ID of the job owner.

describeSigningJobResponse_signingParameters :: Lens' DescribeSigningJobResponse (Maybe (HashMap Text Text))

Map of user-assigned key-value pairs used during signing. These values contain any information that you specified for use in your signing job.

describeSigningJobResponse_httpStatus :: Lens' DescribeSigningJobResponse Int

The response's http status code.

Types

Destination

destination_s3 :: Lens' Destination (Maybe S3Destination)

The S3Destination object.

EncryptionAlgorithmOptions

encryptionAlgorithmOptions_allowedValues :: Lens' EncryptionAlgorithmOptions [EncryptionAlgorithm]

The set of accepted encryption algorithms that are allowed in a code signing job.

encryptionAlgorithmOptions_defaultValue :: Lens' EncryptionAlgorithmOptions EncryptionAlgorithm

The default encryption algorithm that is used by a code signing job.

HashAlgorithmOptions

hashAlgorithmOptions_allowedValues :: Lens' HashAlgorithmOptions [HashAlgorithm]

The set of accepted hash algorithms allowed in a code signing job.

hashAlgorithmOptions_defaultValue :: Lens' HashAlgorithmOptions HashAlgorithm

The default hash algorithm that is used in a code signing job.

Permission

permission_action :: Lens' Permission (Maybe Text)

An AWS Signer action permitted as part of cross-account permissions.

permission_profileVersion :: Lens' Permission (Maybe Text)

The signing profile version that a permission applies to.

permission_principal :: Lens' Permission (Maybe Text)

The AWS principal that has been granted a cross-account permission.

permission_statementId :: Lens' Permission (Maybe Text)

A unique identifier for a cross-account permission statement.

S3Destination

s3Destination_prefix :: Lens' S3Destination (Maybe Text)

An Amazon S3 prefix that you can use to limit responses to those that begin with the specified prefix.

s3Destination_bucketName :: Lens' S3Destination (Maybe Text)

Name of the S3 bucket.

S3SignedObject

s3SignedObject_bucketName :: Lens' S3SignedObject (Maybe Text)

Name of the S3 bucket.

s3SignedObject_key :: Lens' S3SignedObject (Maybe Text)

Key name that uniquely identifies a signed code image in your bucket.

S3Source

s3Source_bucketName :: Lens' S3Source Text

Name of the S3 bucket.

s3Source_key :: Lens' S3Source Text

Key name of the bucket object that contains your unsigned code.

s3Source_version :: Lens' S3Source Text

Version of your source image in your version enabled S3 bucket.

SignatureValidityPeriod

signatureValidityPeriod_value :: Lens' SignatureValidityPeriod (Maybe Int)

The numerical value of the time unit for signature validity.

signatureValidityPeriod_type :: Lens' SignatureValidityPeriod (Maybe ValidityType)

The time unit for signature validity.

SignedObject

signedObject_s3 :: Lens' SignedObject (Maybe S3SignedObject)

The S3SignedObject.

SigningConfiguration

signingConfiguration_encryptionAlgorithmOptions :: Lens' SigningConfiguration EncryptionAlgorithmOptions

The encryption algorithm options that are available for a code signing job.

signingConfiguration_hashAlgorithmOptions :: Lens' SigningConfiguration HashAlgorithmOptions

The hash algorithm options that are available for a code signing job.

SigningConfigurationOverrides

signingConfigurationOverrides_hashAlgorithm :: Lens' SigningConfigurationOverrides (Maybe HashAlgorithm)

A specified override of the default hash algorithm that is used in a code signing job.

signingConfigurationOverrides_encryptionAlgorithm :: Lens' SigningConfigurationOverrides (Maybe EncryptionAlgorithm)

A specified override of the default encryption algorithm that is used in a code signing job.

SigningImageFormat

signingImageFormat_supportedFormats :: Lens' SigningImageFormat [ImageFormat]

The supported formats of a code signing image.

signingImageFormat_defaultFormat :: Lens' SigningImageFormat ImageFormat

The default format of a code signing image.

SigningJob

signingJob_status :: Lens' SigningJob (Maybe SigningStatus)

The status of the signing job.

signingJob_platformDisplayName :: Lens' SigningJob (Maybe Text)

The name of a signing platform.

signingJob_jobId :: Lens' SigningJob (Maybe Text)

The ID of the signing job.

signingJob_createdAt :: Lens' SigningJob (Maybe UTCTime)

The date and time that the signing job was created.

signingJob_signingMaterial :: Lens' SigningJob (Maybe SigningMaterial)

A SigningMaterial object that contains the Amazon Resource Name (ARN) of the certificate used for the signing job.

signingJob_isRevoked :: Lens' SigningJob (Maybe Bool)

Indicates whether the signing job is revoked.

signingJob_signatureExpiresAt :: Lens' SigningJob (Maybe UTCTime)

The time when the signature of a signing job expires.

signingJob_profileVersion :: Lens' SigningJob (Maybe Text)

The version of the signing profile that created a signing job.

signingJob_profileName :: Lens' SigningJob (Maybe Text)

The name of the signing profile that created a signing job.

signingJob_signedObject :: Lens' SigningJob (Maybe SignedObject)

A SignedObject structure that contains information about a signing job's signed code image.

signingJob_platformId :: Lens' SigningJob (Maybe Text)

The unique identifier for a signing platform.

signingJob_source :: Lens' SigningJob (Maybe Source)

A Source that contains information about a signing job's code image source.

signingJob_jobInvoker :: Lens' SigningJob (Maybe Text)

The AWS account ID of the job invoker.

signingJob_jobOwner :: Lens' SigningJob (Maybe Text)

The AWS account ID of the job owner.

SigningJobRevocationRecord

signingJobRevocationRecord_revokedBy :: Lens' SigningJobRevocationRecord (Maybe Text)

The identity of the revoker.

signingJobRevocationRecord_revokedAt :: Lens' SigningJobRevocationRecord (Maybe UTCTime)

The time of revocation.

signingJobRevocationRecord_reason :: Lens' SigningJobRevocationRecord (Maybe Text)

A caller-supplied reason for revocation.

SigningMaterial

signingMaterial_certificateArn :: Lens' SigningMaterial Text

The Amazon Resource Name (ARN) of the certificates that is used to sign your code.

SigningPlatform

signingPlatform_category :: Lens' SigningPlatform (Maybe Category)

The category of a code signing platform.

signingPlatform_signingConfiguration :: Lens' SigningPlatform (Maybe SigningConfiguration)

The configuration of a code signing platform. This includes the designated hash algorithm and encryption algorithm of a signing platform.

signingPlatform_partner :: Lens' SigningPlatform (Maybe Text)

Any partner entities linked to a code signing platform.

signingPlatform_revocationSupported :: Lens' SigningPlatform (Maybe Bool)

Indicates whether revocation is supported for the platform.

signingPlatform_signingImageFormat :: Lens' SigningPlatform (Maybe SigningImageFormat)

Undocumented member.

signingPlatform_platformId :: Lens' SigningPlatform (Maybe Text)

The ID of a code signing; platform.

signingPlatform_displayName :: Lens' SigningPlatform (Maybe Text)

The display name of a code signing platform.

signingPlatform_maxSizeInMB :: Lens' SigningPlatform (Maybe Int)

The maximum size (in MB) of code that can be signed by a code signing platform.

signingPlatform_target :: Lens' SigningPlatform (Maybe Text)

The types of targets that can be signed by a code signing platform.

SigningPlatformOverrides

signingPlatformOverrides_signingConfiguration :: Lens' SigningPlatformOverrides (Maybe SigningConfigurationOverrides)

A signing configuration that overrides the default encryption or hash algorithm of a signing job.

signingPlatformOverrides_signingImageFormat :: Lens' SigningPlatformOverrides (Maybe ImageFormat)

A signed image is a JSON object. When overriding the default signing platform configuration, a customer can select either of two signing formats, JSONEmbedded or JSONDetached. (A third format value, JSON, is reserved for future use.) With JSONEmbedded, the signing image has the payload embedded in it. With JSONDetached, the payload is not be embedded in the signing image.

SigningProfile

signingProfile_status :: Lens' SigningProfile (Maybe SigningProfileStatus)

The status of a code signing profile.

signingProfile_platformDisplayName :: Lens' SigningProfile (Maybe Text)

The name of the signing platform.

signingProfile_arn :: Lens' SigningProfile (Maybe Text)

The Amazon Resource Name (ARN) for the signing profile.

signingProfile_signingMaterial :: Lens' SigningProfile (Maybe SigningMaterial)

The ACM certificate that is available for use by a signing profile.

signingProfile_profileVersion :: Lens' SigningProfile (Maybe Text)

The version of a signing profile.

signingProfile_profileName :: Lens' SigningProfile (Maybe Text)

The name of the signing profile.

signingProfile_profileVersionArn :: Lens' SigningProfile (Maybe Text)

The ARN of a signing profile, including the profile version.

signingProfile_platformId :: Lens' SigningProfile (Maybe Text)

The ID of a platform that is available for use by a signing profile.

signingProfile_signatureValidityPeriod :: Lens' SigningProfile (Maybe SignatureValidityPeriod)

The validity period for a signing job created using this signing profile.

signingProfile_signingParameters :: Lens' SigningProfile (Maybe (HashMap Text Text))

The parameters that are available for use by a code signing user.

signingProfile_tags :: Lens' SigningProfile (Maybe (HashMap Text Text))

A list of tags associated with the signing profile.

SigningProfileRevocationRecord

signingProfileRevocationRecord_revokedBy :: Lens' SigningProfileRevocationRecord (Maybe Text)

The identity of the revoker.

signingProfileRevocationRecord_revocationEffectiveFrom :: Lens' SigningProfileRevocationRecord (Maybe UTCTime)

The time when revocation becomes effective.

signingProfileRevocationRecord_revokedAt :: Lens' SigningProfileRevocationRecord (Maybe UTCTime)

The time when the signing profile was revoked.

Source

source_s3 :: Lens' Source (Maybe S3Source)

The S3Source object.