Network.AWS.Signer.ListSigningJobs

Description

Lists all your signing jobs. You can use the maxResults parameter to limit the number of signing jobs that are returned in the response. If additional jobs remain to be listed, code signing returns a nextToken value. Use this value in subsequent calls to ListSigningJobs to fetch the remaining values. You can continue calling ListSigningJobs with your maxResults parameter and with new values that code signing returns in the nextToken parameter until all of your signing jobs have been returned.

This operation returns paginated results.

Creating a Request

data ListSigningJobs

See: newListSigningJobs smart constructor.

Constructors

ListSigningJobs'

Fields

status :: Maybe SigningStatus

A status value with which to filter your results.

signatureExpiresAfter :: Maybe POSIX

Filters results to return only signing jobs with signatures expiring after a specified timestamp.

requestedBy :: Maybe Text

The IAM principal that requested the signing job.

isRevoked :: Maybe Bool

Filters results to return only signing jobs with revoked signatures.

nextToken :: Maybe Text

String for specifying the next set of paginated results to return. After you receive a response with truncated results, use this parameter in a subsequent request. Set it to the value of nextToken from the response that you just received.

platformId :: Maybe Text

The ID of microcontroller platform that you specified for the distribution of your code image.

jobInvoker :: Maybe Text

Filters results to return only signing jobs initiated by a specified IAM entity.

signatureExpiresBefore :: Maybe POSIX

Filters results to return only signing jobs with signatures expiring before a specified timestamp.

maxResults :: Maybe Natural

Specifies the maximum number of items to return in the response. Use this parameter when paginating results. If additional items exist beyond the number you specify, the nextToken element is set in the response. Use the nextToken value in a subsequent request to retrieve additional items.

newListSigningJobs :: ListSigningJobs

Create a value of ListSigningJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListSigningJobs', listSigningJobs_status - A status value with which to filter your results.

$sel:signatureExpiresAfter:ListSigningJobs', listSigningJobs_signatureExpiresAfter - Filters results to return only signing jobs with signatures expiring after a specified timestamp.

$sel:requestedBy:ListSigningJobs', listSigningJobs_requestedBy - The IAM principal that requested the signing job.

$sel:isRevoked:ListSigningJobs', listSigningJobs_isRevoked - Filters results to return only signing jobs with revoked signatures.

$sel:nextToken:ListSigningJobs', listSigningJobs_nextToken - String for specifying the next set of paginated results to return. After you receive a response with truncated results, use this parameter in a subsequent request. Set it to the value of nextToken from the response that you just received.

$sel:platformId:ListSigningJobs', listSigningJobs_platformId - The ID of microcontroller platform that you specified for the distribution of your code image.

$sel:jobInvoker:ListSigningJobs', listSigningJobs_jobInvoker - Filters results to return only signing jobs initiated by a specified IAM entity.

$sel:signatureExpiresBefore:ListSigningJobs', listSigningJobs_signatureExpiresBefore - Filters results to return only signing jobs with signatures expiring before a specified timestamp.

$sel:maxResults:ListSigningJobs', listSigningJobs_maxResults - Specifies the maximum number of items to return in the response. Use this parameter when paginating results. If additional items exist beyond the number you specify, the nextToken element is set in the response. Use the nextToken value in a subsequent request to retrieve additional items.

Request Lenses

listSigningJobs_status :: Lens' ListSigningJobs (Maybe SigningStatus)

A status value with which to filter your results.

listSigningJobs_signatureExpiresAfter :: Lens' ListSigningJobs (Maybe UTCTime)

Filters results to return only signing jobs with signatures expiring after a specified timestamp.

listSigningJobs_requestedBy :: Lens' ListSigningJobs (Maybe Text)

The IAM principal that requested the signing job.

listSigningJobs_isRevoked :: Lens' ListSigningJobs (Maybe Bool)

Filters results to return only signing jobs with revoked signatures.

listSigningJobs_nextToken :: Lens' ListSigningJobs (Maybe Text)

String for specifying the next set of paginated results to return. After you receive a response with truncated results, use this parameter in a subsequent request. Set it to the value of nextToken from the response that you just received.

listSigningJobs_platformId :: Lens' ListSigningJobs (Maybe Text)

The ID of microcontroller platform that you specified for the distribution of your code image.

listSigningJobs_jobInvoker :: Lens' ListSigningJobs (Maybe Text)

Filters results to return only signing jobs initiated by a specified IAM entity.

listSigningJobs_signatureExpiresBefore :: Lens' ListSigningJobs (Maybe UTCTime)

Filters results to return only signing jobs with signatures expiring before a specified timestamp.

listSigningJobs_maxResults :: Lens' ListSigningJobs (Maybe Natural)

Specifies the maximum number of items to return in the response. Use this parameter when paginating results. If additional items exist beyond the number you specify, the nextToken element is set in the response. Use the nextToken value in a subsequent request to retrieve additional items.

Destructuring the Response

data ListSigningJobsResponse

See: newListSigningJobsResponse smart constructor.

Constructors

ListSigningJobsResponse'

Fields

jobs :: Maybe [SigningJob]

A list of your signing jobs.

nextToken :: Maybe Text

String for specifying the next set of paginated results.

httpStatus :: Int

The response's http status code.

newListSigningJobsResponse

Arguments

:: Int
$sel:httpStatus:ListSigningJobsResponse'

-> ListSigningJobsResponse

Create a value of ListSigningJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobs:ListSigningJobsResponse', listSigningJobsResponse_jobs - A list of your signing jobs.

$sel:nextToken:ListSigningJobs', listSigningJobsResponse_nextToken - String for specifying the next set of paginated results.

$sel:httpStatus:ListSigningJobsResponse', listSigningJobsResponse_httpStatus - The response's http status code.

Response Lenses

listSigningJobsResponse_jobs :: Lens' ListSigningJobsResponse (Maybe [SigningJob])

A list of your signing jobs.

listSigningJobsResponse_nextToken :: Lens' ListSigningJobsResponse (Maybe Text)

String for specifying the next set of paginated results.

listSigningJobsResponse_httpStatus :: Lens' ListSigningJobsResponse Int

The response's http status code.