Network.AWS.Signer.ListSigningPlatforms

Description

Lists all signing platforms available in code signing that match the request parameters. If additional jobs remain to be listed, code signing returns a nextToken value. Use this value in subsequent calls to ListSigningJobs to fetch the remaining values. You can continue calling ListSigningJobs with your maxResults parameter and with new values that code signing returns in the nextToken parameter until all of your signing jobs have been returned.

This operation returns paginated results.

Creating a Request

data ListSigningPlatforms

See: newListSigningPlatforms smart constructor.

Constructors

ListSigningPlatforms'

Fields

category :: Maybe Text

The category type of a signing platform.

partner :: Maybe Text

Any partner entities connected to a signing platform.

nextToken :: Maybe Text

Value for specifying the next set of paginated results to return. After you receive a response with truncated results, use this parameter in a subsequent request. Set it to the value of nextToken from the response that you just received.

maxResults :: Maybe Natural

The maximum number of results to be returned by this operation.

target :: Maybe Text

The validation template that is used by the target signing platform.

newListSigningPlatforms :: ListSigningPlatforms

Create a value of ListSigningPlatforms with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:category:ListSigningPlatforms', listSigningPlatforms_category - The category type of a signing platform.

$sel:partner:ListSigningPlatforms', listSigningPlatforms_partner - Any partner entities connected to a signing platform.

$sel:nextToken:ListSigningPlatforms', listSigningPlatforms_nextToken - Value for specifying the next set of paginated results to return. After you receive a response with truncated results, use this parameter in a subsequent request. Set it to the value of nextToken from the response that you just received.

$sel:maxResults:ListSigningPlatforms', listSigningPlatforms_maxResults - The maximum number of results to be returned by this operation.

$sel:target:ListSigningPlatforms', listSigningPlatforms_target - The validation template that is used by the target signing platform.

Request Lenses

listSigningPlatforms_category :: Lens' ListSigningPlatforms (Maybe Text)

The category type of a signing platform.

listSigningPlatforms_partner :: Lens' ListSigningPlatforms (Maybe Text)

Any partner entities connected to a signing platform.

listSigningPlatforms_nextToken :: Lens' ListSigningPlatforms (Maybe Text)

Value for specifying the next set of paginated results to return. After you receive a response with truncated results, use this parameter in a subsequent request. Set it to the value of nextToken from the response that you just received.

listSigningPlatforms_maxResults :: Lens' ListSigningPlatforms (Maybe Natural)

The maximum number of results to be returned by this operation.

listSigningPlatforms_target :: Lens' ListSigningPlatforms (Maybe Text)

The validation template that is used by the target signing platform.

Destructuring the Response

data ListSigningPlatformsResponse

See: newListSigningPlatformsResponse smart constructor.

Constructors

ListSigningPlatformsResponse'

Fields

platforms :: Maybe [SigningPlatform]

A list of all platforms that match the request parameters.

nextToken :: Maybe Text

Value for specifying the next set of paginated results to return.

httpStatus :: Int

The response's http status code.

newListSigningPlatformsResponse

Arguments

:: Int
$sel:httpStatus:ListSigningPlatformsResponse'

-> ListSigningPlatformsResponse

Create a value of ListSigningPlatformsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platforms:ListSigningPlatformsResponse', listSigningPlatformsResponse_platforms - A list of all platforms that match the request parameters.

$sel:nextToken:ListSigningPlatforms', listSigningPlatformsResponse_nextToken - Value for specifying the next set of paginated results to return.

$sel:httpStatus:ListSigningPlatformsResponse', listSigningPlatformsResponse_httpStatus - The response's http status code.

Response Lenses

listSigningPlatformsResponse_platforms :: Lens' ListSigningPlatformsResponse (Maybe [SigningPlatform])

A list of all platforms that match the request parameters.

listSigningPlatformsResponse_nextToken :: Lens' ListSigningPlatformsResponse (Maybe Text)

Value for specifying the next set of paginated results to return.

listSigningPlatformsResponse_httpStatus :: Lens' ListSigningPlatformsResponse Int

The response's http status code.