Network.AWS.Signer.RemoveProfilePermission

Description

Removes cross-account permissions from a signing profile.

Creating a Request

data RemoveProfilePermission

See: newRemoveProfilePermission smart constructor.

Constructors

RemoveProfilePermission'

Fields

revisionId :: Text

An identifier for the current revision of the signing profile permissions.

profileName :: Text

A human-readable name for the signing profile with permissions to be removed.

statementId :: Text

A unique identifier for the cross-account permissions statement.

newRemoveProfilePermission

Arguments

:: Text
$sel:revisionId:RemoveProfilePermission'

-> Text
$sel:profileName:RemoveProfilePermission'

-> Text
$sel:statementId:RemoveProfilePermission'

-> RemoveProfilePermission

Create a value of RemoveProfilePermission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revisionId:RemoveProfilePermission', removeProfilePermission_revisionId - An identifier for the current revision of the signing profile permissions.

$sel:profileName:RemoveProfilePermission', removeProfilePermission_profileName - A human-readable name for the signing profile with permissions to be removed.

$sel:statementId:RemoveProfilePermission', removeProfilePermission_statementId - A unique identifier for the cross-account permissions statement.

Request Lenses

removeProfilePermission_revisionId :: Lens' RemoveProfilePermission Text

An identifier for the current revision of the signing profile permissions.

removeProfilePermission_profileName :: Lens' RemoveProfilePermission Text

A human-readable name for the signing profile with permissions to be removed.

removeProfilePermission_statementId :: Lens' RemoveProfilePermission Text

A unique identifier for the cross-account permissions statement.

Destructuring the Response

data RemoveProfilePermissionResponse

See: newRemoveProfilePermissionResponse smart constructor.

Constructors

RemoveProfilePermissionResponse'

Fields

revisionId :: Maybe Text

An identifier for the current revision of the profile permissions.

httpStatus :: Int

The response's http status code.

newRemoveProfilePermissionResponse

Arguments

:: Int
$sel:httpStatus:RemoveProfilePermissionResponse'

-> RemoveProfilePermissionResponse

Create a value of RemoveProfilePermissionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revisionId:RemoveProfilePermission', removeProfilePermissionResponse_revisionId - An identifier for the current revision of the profile permissions.

$sel:httpStatus:RemoveProfilePermissionResponse', removeProfilePermissionResponse_httpStatus - The response's http status code.

Response Lenses

removeProfilePermissionResponse_revisionId :: Lens' RemoveProfilePermissionResponse (Maybe Text)

An identifier for the current revision of the profile permissions.

removeProfilePermissionResponse_httpStatus :: Lens' RemoveProfilePermissionResponse Int

The response's http status code.