Network.AWS.Signer.Types.Permission

Description

Documentation

data Permission

A cross-account permission for a signing profile.

See: newPermission smart constructor.

Constructors

Permission'

Fields

action :: Maybe Text

An AWS Signer action permitted as part of cross-account permissions.

profileVersion :: Maybe Text

The signing profile version that a permission applies to.

principal :: Maybe Text

The AWS principal that has been granted a cross-account permission.

statementId :: Maybe Text

A unique identifier for a cross-account permission statement.

newPermission :: Permission

Create a value of Permission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:action:Permission', permission_action - An AWS Signer action permitted as part of cross-account permissions.

$sel:profileVersion:Permission', permission_profileVersion - The signing profile version that a permission applies to.

$sel:principal:Permission', permission_principal - The AWS principal that has been granted a cross-account permission.

$sel:statementId:Permission', permission_statementId - A unique identifier for a cross-account permission statement.

permission_action :: Lens' Permission (Maybe Text)

An AWS Signer action permitted as part of cross-account permissions.

permission_profileVersion :: Lens' Permission (Maybe Text)

The signing profile version that a permission applies to.

permission_principal :: Lens' Permission (Maybe Text)

The AWS principal that has been granted a cross-account permission.

permission_statementId :: Lens' Permission (Maybe Text)

A unique identifier for a cross-account permission statement.