Network.AWS.Signer.Types.S3Source

Description

Documentation

data S3Source

Information about the S3 bucket where you saved your unsigned code.

See: newS3Source smart constructor.

Constructors

S3Source'

Fields

bucketName :: Text

Name of the S3 bucket.

key :: Text

Key name of the bucket object that contains your unsigned code.

version :: Text

Version of your source image in your version enabled S3 bucket.

newS3Source

Arguments

:: Text
$sel:bucketName:S3Source'

-> Text
$sel:key:S3Source'

-> Text
$sel:version:S3Source'

-> S3Source

Create a value of S3Source with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucketName:S3Source', s3Source_bucketName - Name of the S3 bucket.

$sel:key:S3Source', s3Source_key - Key name of the bucket object that contains your unsigned code.

$sel:version:S3Source', s3Source_version - Version of your source image in your version enabled S3 bucket.

s3Source_bucketName :: Lens' S3Source Text

Name of the S3 bucket.

s3Source_key :: Lens' S3Source Text

Key name of the bucket object that contains your unsigned code.

s3Source_version :: Lens' S3Source Text

Version of your source image in your version enabled S3 bucket.