Network.AWS.Signer.Types.SigningProfile

Description

Documentation

data SigningProfile

Contains information about the ACM certificates and code signing configuration parameters that can be used by a given code signing user.

See: newSigningProfile smart constructor.

Constructors

SigningProfile'

Fields

status :: Maybe SigningProfileStatus

The status of a code signing profile.

platformDisplayName :: Maybe Text

The name of the signing platform.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the signing profile.

signingMaterial :: Maybe SigningMaterial

The ACM certificate that is available for use by a signing profile.

profileVersion :: Maybe Text

The version of a signing profile.

profileName :: Maybe Text

The name of the signing profile.

profileVersionArn :: Maybe Text

The ARN of a signing profile, including the profile version.

platformId :: Maybe Text

The ID of a platform that is available for use by a signing profile.

signatureValidityPeriod :: Maybe SignatureValidityPeriod

The validity period for a signing job created using this signing profile.

signingParameters :: Maybe (HashMap Text Text)

The parameters that are available for use by a code signing user.

tags :: Maybe (HashMap Text Text)

A list of tags associated with the signing profile.

newSigningProfile :: SigningProfile

Create a value of SigningProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SigningProfile', signingProfile_status - The status of a code signing profile.

$sel:platformDisplayName:SigningProfile', signingProfile_platformDisplayName - The name of the signing platform.

$sel:arn:SigningProfile', signingProfile_arn - The Amazon Resource Name (ARN) for the signing profile.

$sel:signingMaterial:SigningProfile', signingProfile_signingMaterial - The ACM certificate that is available for use by a signing profile.

$sel:profileVersion:SigningProfile', signingProfile_profileVersion - The version of a signing profile.

$sel:profileName:SigningProfile', signingProfile_profileName - The name of the signing profile.

$sel:profileVersionArn:SigningProfile', signingProfile_profileVersionArn - The ARN of a signing profile, including the profile version.

$sel:platformId:SigningProfile', signingProfile_platformId - The ID of a platform that is available for use by a signing profile.

$sel:signatureValidityPeriod:SigningProfile', signingProfile_signatureValidityPeriod - The validity period for a signing job created using this signing profile.

$sel:signingParameters:SigningProfile', signingProfile_signingParameters - The parameters that are available for use by a code signing user.

$sel:tags:SigningProfile', signingProfile_tags - A list of tags associated with the signing profile.

signingProfile_status :: Lens' SigningProfile (Maybe SigningProfileStatus)

The status of a code signing profile.

signingProfile_platformDisplayName :: Lens' SigningProfile (Maybe Text)

The name of the signing platform.

signingProfile_arn :: Lens' SigningProfile (Maybe Text)

The Amazon Resource Name (ARN) for the signing profile.

signingProfile_signingMaterial :: Lens' SigningProfile (Maybe SigningMaterial)

The ACM certificate that is available for use by a signing profile.

signingProfile_profileVersion :: Lens' SigningProfile (Maybe Text)

The version of a signing profile.

signingProfile_profileName :: Lens' SigningProfile (Maybe Text)

The name of the signing profile.

signingProfile_profileVersionArn :: Lens' SigningProfile (Maybe Text)

The ARN of a signing profile, including the profile version.

signingProfile_platformId :: Lens' SigningProfile (Maybe Text)

The ID of a platform that is available for use by a signing profile.

signingProfile_signatureValidityPeriod :: Lens' SigningProfile (Maybe SignatureValidityPeriod)

The validity period for a signing job created using this signing profile.

signingProfile_signingParameters :: Lens' SigningProfile (Maybe (HashMap Text Text))

The parameters that are available for use by a code signing user.

signingProfile_tags :: Lens' SigningProfile (Maybe (HashMap Text Text))

A list of tags associated with the signing profile.