Network.AWS.SimpleDB.V2009_04_15

Description

Amazon SimpleDB is a highly available and flexible non-relational data store that offloads the work of database administration. Developers simply store and query data items via web services requests and Amazon SimpleDB does the rest. Unbound by the strict requirements of a relational database, Amazon SimpleDB is optimized to provide high availability and flexibility, with little or no administrative burden. Behind the scenes, Amazon SimpleDB creates and manages multiple geographically distributed replicas of your data automatically to enable high availability and data durability. The service charges you only for the resources actually consumed in storing your data and serving your requests. You can change your data model on the fly, and data is automatically indexed for you. With Amazon SimpleDB, you can focus on application development without worrying about infrastructure provisioning, high availability, software maintenance, schema and index management, or performance tuning.

Documentation

data BatchDeleteAttributes

Constructors

BatchDeleteAttributes

Fields

_bdarItems :: [DeletableItem]

A list of items on which to perform the operation.

_bdarDomainName :: Text

The name of the domain in which the attributes are being deleted.

bdarItems :: Lens' BatchDeleteAttributes [DeletableItem]

bdarDomainName :: Lens' BatchDeleteAttributes Text

data BatchDeleteAttributesResponse

Constructors

BatchDeleteAttributesResponse

data BatchPutAttributes

Constructors

BatchPutAttributes

Fields

_bparItems :: [ReplaceableItem]

A list of items on which to perform the operation.

_bparDomainName :: Text

The name of the domain in which the attributes are being stored.

bparItems :: Lens' BatchPutAttributes [ReplaceableItem]

bparDomainName :: Lens' BatchPutAttributes Text

data BatchPutAttributesResponse

Constructors

BatchPutAttributesResponse

data CreateDomain

Constructors

CreateDomain

Fields

_cdrDomainName :: Text

The name of the domain to create. The name can range between 3 and 255 characters and can contain the following characters: a-z, A-Z, 0-9, '_', -, and ..

cdrDomainName :: Iso' CreateDomain Text

data CreateDomainResponse

Constructors

CreateDomainResponse

deleteAttributes

Arguments

:: Text
_darItemName

-> Text
_darDomainName

-> DeleteAttributes

Minimum specification for a DeleteAttributes request.

data DeleteAttributes

Constructors

DeleteAttributes

Fields

_darItemName :: Text

The name of the item. Similar to rows on a spreadsheet, items represent individual objects that contain one or more value-attribute pairs.

_darDomainName :: Text

The name of the domain in which to perform the operation.

_darAttributes :: [Attribute]

A list of Attributes. Similar to columns on a spreadsheet, attributes represent categories of data that can be assigned to items.

_darExpected :: Maybe UpdateCondition

The update condition which, if specified, determines whether the specified attributes will be deleted or not. The update condition must be satisfied in order for this request to be processed and the attributes to be deleted.

darItemName :: Lens' DeleteAttributes Text

darExpected :: Lens' DeleteAttributes (Maybe UpdateCondition)

darDomainName :: Lens' DeleteAttributes Text

darAttributes :: Lens' DeleteAttributes [Attribute]

data DeleteAttributesResponse

Constructors

DeleteAttributesResponse

data DeleteDomain

Constructors

DeleteDomain

Fields

_ddrDomainName :: Text

The name of the domain to delete.

ddrDomainName :: Iso' DeleteDomain Text

data DeleteDomainResponse

Constructors

DeleteDomainResponse

data DomainMetadata

Constructors

DomainMetadata

Fields

_dmrDomainName :: Text

The name of the domain for which to display the metadata of.

dmrDomainName :: Iso' DomainMetadata Text

data DomainMetadataResponse

Constructors

DomainMetadataResponse

Fields

_dmsAttributeNameCount :: Maybe Integer

The number of unique attribute names in the domain.

_dmsAttributeValueCount :: Maybe Integer

The number of all attribute name/value pairs in the domain.

_dmsItemCount :: Maybe Integer

The number of all items in the domain.

_dmsTimestamp :: Maybe Integer

The data and time when metadata was calculated, in Epoch (UNIX) seconds.

_dmsItemNamesSizeBytes :: Maybe Integer

The total size of all item names in the domain, in bytes.

_dmsAttributeValuesSizeBytes :: Maybe Integer

The total size of all attribute values in the domain, in bytes.

_dmsAttributeNamesSizeBytes :: Maybe Integer

The total size of all unique attribute names in the domain, in bytes.

dmsTimestamp :: Lens' DomainMetadataResponse (Maybe Integer)

dmsItemNamesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

dmsItemCount :: Lens' DomainMetadataResponse (Maybe Integer)

dmsAttributeValuesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

dmsAttributeValueCount :: Lens' DomainMetadataResponse (Maybe Integer)

dmsAttributeNamesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

dmsAttributeNameCount :: Lens' DomainMetadataResponse (Maybe Integer)

getAttributes

Arguments

:: Text
_garItemName

-> Text
_garDomainName

-> GetAttributes

Minimum specification for a GetAttributes request.

data GetAttributes

Constructors

GetAttributes

Fields

_garItemName :: Text

The name of the item.

_garDomainName :: Text

The name of the domain in which to perform the operation.

_garAttributeNames :: [Text]

The names of the attributes.

_garConsistentRead :: Maybe Bool

Determines whether or not strong consistency should be enforced when data is read from SimpleDB. If true, any data previously written to SimpleDB will be returned. Otherwise, results will be consistent eventually, and the client may not see data that was written immediately before your read.

garItemName :: Lens' GetAttributes Text

garDomainName :: Lens' GetAttributes Text

garConsistentRead :: Lens' GetAttributes (Maybe Bool)

garAttributeNames :: Lens' GetAttributes [Text]

data GetAttributesResponse

Constructors

GetAttributesResponse

Fields

_gasAttributes :: [Attribute]

The list of attributes returned by the operation.

gasAttributes :: Iso' GetAttributesResponse [Attribute]

listDomains :: ListDomains

Minimum specification for a ListDomains request.

data ListDomains

Constructors

ListDomains

Fields

_ldrMaxNumberOfDomains :: Maybe Integer

The maximum number of domain names you want returned. The range is 1 to 100. The default setting is 100.

_ldrNextToken :: Maybe Text

A string informing Amazon SimpleDB where to start the next list of domain names.

ldrNextToken :: Lens' ListDomains (Maybe Text)

ldrMaxNumberOfDomains :: Lens' ListDomains (Maybe Integer)

data ListDomainsResponse

Constructors

ListDomainsResponse

Fields

_ldsDomainNames :: [Text]

A list of domain names that match the expression.

_ldsNextToken :: Maybe Text

An opaque token indicating that there are more domains than the specified MaxNumberOfDomains still available.

ldsNextToken :: Lens' ListDomainsResponse (Maybe Text)

ldsDomainNames :: Lens' ListDomainsResponse [Text]

putAttributes

Arguments

:: [ReplaceableAttribute]
_parAttributes

-> Text
_parItemName

-> Text
_parDomainName

-> PutAttributes

Minimum specification for a PutAttributes request.

data PutAttributes

Constructors

PutAttributes

Fields

_parAttributes :: [ReplaceableAttribute]

The list of attributes.

_parItemName :: Text

The name of the item.

_parDomainName :: Text

The name of the domain in which to perform the operation.

_parExpected :: Maybe UpdateCondition

The update condition which, if specified, determines whether the specified attributes will be updated or not. The update condition must be satisfied in order for this request to be processed and the attributes to be updated.

parItemName :: Lens' PutAttributes Text

parExpected :: Lens' PutAttributes (Maybe UpdateCondition)

parDomainName :: Lens' PutAttributes Text

parAttributes :: Lens' PutAttributes [ReplaceableAttribute]

data PutAttributesResponse

Constructors

PutAttributesResponse

select

Arguments

:: Text
_srSelectExpression

-> Select

Minimum specification for a Select request.

data Select

Constructors

Select

Fields

_srSelectExpression :: Text

The expression used to query the domain.

_srConsistentRead :: Maybe Bool

Determines whether or not strong consistency should be enforced when data is read from SimpleDB. If true, any data previously written to SimpleDB will be returned. Otherwise, results will be consistent eventually, and the client may not see data that was written immediately before your read.

_srNextToken :: Maybe Text

A string informing Amazon SimpleDB where to start the next list of ItemNames.

srSelectExpression :: Lens' Select Text

srNextToken :: Lens' Select (Maybe Text)

srConsistentRead :: Lens' Select (Maybe Bool)

data SelectResponse

Constructors

SelectResponse

Fields

_suItems :: [Item]

A list of items that match the select expression.

_suNextToken :: Maybe Text

An opaque token indicating that more items than MaxNumberOfItems were matched, the response size exceeded 1 megabyte, or the execution time exceeded 5 seconds.

suNextToken :: Lens' SelectResponse (Maybe Text)

suItems :: Lens' SelectResponse [Item]

data SimpleDB

Supported version (2009-04-15) of the Amazon SimpleDB service.

xmlOptions :: Tagged a XMLOptions

data Attribute

Constructors

Attribute

Fields

_aAlternateValueEncoding :: Maybe Text

_aValue :: Text

The value of the attribute.

_aAlternateNameEncoding :: Maybe Text

_aName :: Text

The name of the attribute.

data DeletableItem

Constructors

DeletableItem

Fields

_diAttributes :: [Attribute]

_diName :: Text

data Item

Constructors

Item

Fields

_pAlternateNameEncoding :: Maybe Text

_pAttributes :: [Attribute]

A list of attributes.

_pName :: Text

The name of the item.

data ReplaceableAttribute

Constructors

ReplaceableAttribute

Fields

_raReplace :: Maybe Bool

A flag specifying whether or not to replace the attribute/value pair or to add a new attribute/value pair. The default setting is false.

_raValue :: Text

The value of the replaceable attribute.

_raName :: Text

The name of the replaceable attribute.

data ReplaceableItem

Constructors

ReplaceableItem

Fields

_riAttributes :: [ReplaceableAttribute]

The list of attributes for a replaceable item.

_riName :: Text

The name of the replaceable item.

data UpdateCondition

The update condition which, if specified, determines whether the specified attributes will be deleted or not. The update condition must be satisfied in order for this request to be processed and the attributes to be deleted.

Constructors

UpdateCondition

Fields

_ucExists :: Maybe Bool

A value specifying whether or not the specified attribute must exist with the specified value in order for the update condition to be satisfied. Specify true if the attribute must exist for the update condition to be satisfied. Specify false if the attribute should not exist in order for the update condition to be satisfied.

_ucValue :: Maybe Text

The value of an attribute. This value can only be specified when the Exists parameter is equal to true.

_ucName :: Maybe Text

The name of the attribute involved in the condition.

aValue :: Lens' Attribute Text

aName :: Lens' Attribute Text

aAlternateValueEncoding :: Lens' Attribute (Maybe Text)

aAlternateNameEncoding :: Lens' Attribute (Maybe Text)

diName :: Lens' DeletableItem Text

diAttributes :: Lens' DeletableItem [Attribute]

pName :: Lens' Item Text

pAttributes :: Lens' Item [Attribute]

pAlternateNameEncoding :: Lens' Item (Maybe Text)

raValue :: Lens' ReplaceableAttribute Text

raReplace :: Lens' ReplaceableAttribute (Maybe Bool)

raName :: Lens' ReplaceableAttribute Text

riName :: Lens' ReplaceableItem Text

riAttributes :: Lens' ReplaceableItem [ReplaceableAttribute]

ucValue :: Lens' UpdateCondition (Maybe Text)

ucName :: Lens' UpdateCondition (Maybe Text)

ucExists :: Lens' UpdateCondition (Maybe Bool)