Network.AWS.SimpleDB.V2009_04_15.CreateDomain

Description

The CreateDomain operation creates a new domain. The domain name should be unique among the domains associated with the Access Key ID provided in the request. The CreateDomain operation may take 10 or more seconds to complete. CreateDomain is an idempotent operation; running it multiple times using the same domain name will not result in an error response. The client can create up to 100 domains per account. If the client requires additional domains, go to http:/aws.amazon.comcontact-ussimpledb-limit-request.

Documentation

data CreateDomain

Constructors

CreateDomain

Fields

_cdrDomainName :: Text

The name of the domain to create. The name can range between 3 and 255 characters and can contain the following characters: a-z, A-Z, 0-9, '_', -, and ..

cdrDomainName :: Iso' CreateDomain Text

data CreateDomainResponse

Constructors

CreateDomainResponse