Network.AWS.SimpleDB.V2009_04_15.Types

Description

Amazon SimpleDB is a highly available and flexible non-relational data store that offloads the work of database administration. Developers simply store and query data items via web services requests and Amazon SimpleDB does the rest. Unbound by the strict requirements of a relational database, Amazon SimpleDB is optimized to provide high availability and flexibility, with little or no administrative burden. Behind the scenes, Amazon SimpleDB creates and manages multiple geographically distributed replicas of your data automatically to enable high availability and data durability. The service charges you only for the resources actually consumed in storing your data and serving your requests. You can change your data model on the fly, and data is automatically indexed for you. With Amazon SimpleDB, you can focus on application development without worrying about infrastructure provisioning, high availability, software maintenance, schema and index management, or performance tuning.

Documentation

data SimpleDB

Supported version (2009-04-15) of the Amazon SimpleDB service.

xmlOptions :: Tagged a XMLOptions

data Attribute

Constructors

Attribute

Fields

_aAlternateValueEncoding :: Maybe Text

_aValue :: Text

The value of the attribute.

_aAlternateNameEncoding :: Maybe Text

_aName :: Text

The name of the attribute.

data DeletableItem

Constructors

DeletableItem

Fields

_diAttributes :: [Attribute]

_diName :: Text

data Item

Constructors

Item

Fields

_pAlternateNameEncoding :: Maybe Text

_pAttributes :: [Attribute]

A list of attributes.

_pName :: Text

The name of the item.

data ReplaceableAttribute

Constructors

ReplaceableAttribute

Fields

_raReplace :: Maybe Bool

A flag specifying whether or not to replace the attribute/value pair or to add a new attribute/value pair. The default setting is false.

_raValue :: Text

The value of the replaceable attribute.

_raName :: Text

The name of the replaceable attribute.

data ReplaceableItem

Constructors

ReplaceableItem

Fields

_riAttributes :: [ReplaceableAttribute]

The list of attributes for a replaceable item.

_riName :: Text

The name of the replaceable item.

data UpdateCondition

The update condition which, if specified, determines whether the specified attributes will be deleted or not. The update condition must be satisfied in order for this request to be processed and the attributes to be deleted.

Constructors

UpdateCondition

Fields

_ucExists :: Maybe Bool

A value specifying whether or not the specified attribute must exist with the specified value in order for the update condition to be satisfied. Specify true if the attribute must exist for the update condition to be satisfied. Specify false if the attribute should not exist in order for the update condition to be satisfied.

_ucValue :: Maybe Text

The value of an attribute. This value can only be specified when the Exists parameter is equal to true.

_ucName :: Maybe Text

The name of the attribute involved in the condition.

aValue :: Lens' Attribute Text

aName :: Lens' Attribute Text

aAlternateValueEncoding :: Lens' Attribute (Maybe Text)

aAlternateNameEncoding :: Lens' Attribute (Maybe Text)

diName :: Lens' DeletableItem Text

diAttributes :: Lens' DeletableItem [Attribute]

pName :: Lens' Item Text

pAttributes :: Lens' Item [Attribute]

pAlternateNameEncoding :: Lens' Item (Maybe Text)

raValue :: Lens' ReplaceableAttribute Text

raReplace :: Lens' ReplaceableAttribute (Maybe Bool)

raName :: Lens' ReplaceableAttribute Text

riName :: Lens' ReplaceableItem Text

riAttributes :: Lens' ReplaceableItem [ReplaceableAttribute]

ucValue :: Lens' UpdateCondition (Maybe Text)

ucName :: Lens' UpdateCondition (Maybe Text)

ucExists :: Lens' UpdateCondition (Maybe Bool)