Network.AWS.SnowDeviceManagement

Description

Derived from API version 2021-08-04 of the AWS service descriptions, licensed under Apache 2.0.

Amazon Web Services Snow Device Management documentation.

Service Configuration

defaultService :: Service

API version 2021-08-04 of the Amazon Snow Device Management SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by SnowDeviceManagement.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by an Amazon Web Services service.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You don't have sufficient access to perform this action.

ServiceQuotaExceededException

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request would cause a service quota to be exceeded.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

An unexpected error occurred while processing the request.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The request references a resource that doesn't exist.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The Amazon Resource Name (ARN) of the device or task.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - The list of tags for the device or task.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

DescribeTask

data DescribeTask

See: newDescribeTask smart constructor.

Constructors

DescribeTask' Text

newDescribeTask

Arguments

:: Text
$sel:taskId:DescribeTask'

-> DescribeTask

Create a value of DescribeTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskId:DescribeTask', describeTask_taskId - The ID of the task to be described.

data DescribeTaskResponse

See: newDescribeTaskResponse smart constructor.

Constructors

DescribeTaskResponse' (Maybe TaskState) (Maybe POSIX) (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe (NonEmpty Text)) (Maybe POSIX) (Maybe Text) (Maybe (HashMap Text Text)) Int

newDescribeTaskResponse

Arguments

:: Int
$sel:httpStatus:DescribeTaskResponse'

-> DescribeTaskResponse

Create a value of DescribeTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:DescribeTaskResponse', describeTaskResponse_state - The current state of the task.

$sel:lastUpdatedAt:DescribeTaskResponse', describeTaskResponse_lastUpdatedAt - When the state of the task was last updated.

$sel:createdAt:DescribeTaskResponse', describeTaskResponse_createdAt - When the CreateTask operation was called.

$sel:taskId:DescribeTask', describeTaskResponse_taskId - The ID of the task.

$sel:taskArn:DescribeTaskResponse', describeTaskResponse_taskArn - The Amazon Resource Name (ARN) of the task.

$sel:targets:DescribeTaskResponse', describeTaskResponse_targets - The managed devices that the task was sent to.

$sel:completedAt:DescribeTaskResponse', describeTaskResponse_completedAt - When the task was completed.

$sel:description:DescribeTaskResponse', describeTaskResponse_description - The description provided of the task and managed devices.

$sel:tags:DescribeTaskResponse', describeTaskResponse_tags - Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

$sel:httpStatus:DescribeTaskResponse', describeTaskResponse_httpStatus - The response's http status code.

ListDeviceResources (Paginated)

data ListDeviceResources

See: newListDeviceResources smart constructor.

Constructors

ListDeviceResources' (Maybe Text) (Maybe Text) (Maybe Natural) Text

newListDeviceResources

Arguments

:: Text
$sel:managedDeviceId:ListDeviceResources'

-> ListDeviceResources

Create a value of ListDeviceResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDeviceResources', listDeviceResources_nextToken - A pagination token to continue to the next page of results.

$sel:type':ListDeviceResources', listDeviceResources_type - A structure used to filter the results by type of resource.

$sel:maxResults:ListDeviceResources', listDeviceResources_maxResults - The maximum number of resources per page.

$sel:managedDeviceId:ListDeviceResources', listDeviceResources_managedDeviceId - The ID of the managed device that you are listing the resources of.

data ListDeviceResourcesResponse

See: newListDeviceResourcesResponse smart constructor.

Constructors

ListDeviceResourcesResponse' (Maybe [ResourceSummary]) (Maybe Text) Int

newListDeviceResourcesResponse

Arguments

:: Int
$sel:httpStatus:ListDeviceResourcesResponse'

-> ListDeviceResourcesResponse

Create a value of ListDeviceResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resources:ListDeviceResourcesResponse', listDeviceResourcesResponse_resources - A structure defining the resource's type, Amazon Resource Name (ARN), and ID.

$sel:nextToken:ListDeviceResources', listDeviceResourcesResponse_nextToken - A pagination token to continue to the next page of results.

$sel:httpStatus:ListDeviceResourcesResponse', listDeviceResourcesResponse_httpStatus - The response's http status code.

ListExecutions (Paginated)

data ListExecutions

See: newListExecutions smart constructor.

Constructors

ListExecutions' (Maybe ExecutionState) (Maybe Text) (Maybe Natural) Text

newListExecutions

Arguments

:: Text
$sel:taskId:ListExecutions'

-> ListExecutions

Create a value of ListExecutions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:ListExecutions', listExecutions_state - A structure used to filter the tasks by their current state.

$sel:nextToken:ListExecutions', listExecutions_nextToken - A pagination token to continue to the next page of tasks.

$sel:maxResults:ListExecutions', listExecutions_maxResults - The maximum number of tasks to list per page.

$sel:taskId:ListExecutions', listExecutions_taskId - The ID of the task.

data ListExecutionsResponse

See: newListExecutionsResponse smart constructor.

Constructors

ListExecutionsResponse' (Maybe [ExecutionSummary]) (Maybe Text) Int

newListExecutionsResponse

Arguments

:: Int
$sel:httpStatus:ListExecutionsResponse'

-> ListExecutionsResponse

Create a value of ListExecutionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:executions:ListExecutionsResponse', listExecutionsResponse_executions - A list of executions. Each execution contains the task ID, the device that the task is executing on, the execution ID, and the status of the execution.

$sel:nextToken:ListExecutions', listExecutionsResponse_nextToken - A pagination token to continue to the next page of executions.

$sel:httpStatus:ListExecutionsResponse', listExecutionsResponse_httpStatus - The response's http status code.

DescribeDeviceEc2Instances

data DescribeDeviceEc2Instances

See: newDescribeDeviceEc2Instances smart constructor.

Constructors

DescribeDeviceEc2Instances' [Text] Text

newDescribeDeviceEc2Instances

Arguments

:: Text
$sel:managedDeviceId:DescribeDeviceEc2Instances'

-> DescribeDeviceEc2Instances

Create a value of DescribeDeviceEc2Instances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceIds:DescribeDeviceEc2Instances', describeDeviceEc2Instances_instanceIds - A list of instance IDs associated with the managed device.

$sel:managedDeviceId:DescribeDeviceEc2Instances', describeDeviceEc2Instances_managedDeviceId - The ID of the managed device.

data DescribeDeviceEc2InstancesResponse

See: newDescribeDeviceEc2InstancesResponse smart constructor.

Constructors

DescribeDeviceEc2InstancesResponse' (Maybe [InstanceSummary]) Int

newDescribeDeviceEc2InstancesResponse

Arguments

:: Int
$sel:httpStatus:DescribeDeviceEc2InstancesResponse'

-> DescribeDeviceEc2InstancesResponse

Create a value of DescribeDeviceEc2InstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instances:DescribeDeviceEc2InstancesResponse', describeDeviceEc2InstancesResponse_instances - A list of structures containing information about each instance.

$sel:httpStatus:DescribeDeviceEc2InstancesResponse', describeDeviceEc2InstancesResponse_httpStatus - The response's http status code.

CreateTask

data CreateTask

See: newCreateTask smart constructor.

Constructors

CreateTask' (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) Command (NonEmpty Text)

newCreateTask

Arguments

:: Command
$sel:command:CreateTask'

-> NonEmpty Text
$sel:targets:CreateTask'

-> CreateTask

Create a value of CreateTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateTask', createTask_clientToken - A token ensuring that the action is called only once with the specified details.

$sel:description:CreateTask', createTask_description - A description of the task and its targets.

$sel:tags:CreateTask', createTask_tags - Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

$sel:command:CreateTask', createTask_command - The task to be performed. Only one task is executed on a device at a time.

$sel:targets:CreateTask', createTask_targets - A list of managed device IDs.

data CreateTaskResponse

See: newCreateTaskResponse smart constructor.

Constructors

CreateTaskResponse' (Maybe Text) (Maybe Text) Int

newCreateTaskResponse

Arguments

:: Int
$sel:httpStatus:CreateTaskResponse'

-> CreateTaskResponse

Create a value of CreateTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskId:CreateTaskResponse', createTaskResponse_taskId - The ID of the task that you created.

$sel:taskArn:CreateTaskResponse', createTaskResponse_taskArn - The Amazon Resource Name (ARN) of the task that you created.

$sel:httpStatus:CreateTaskResponse', createTaskResponse_httpStatus - The response's http status code.

ListTasks (Paginated)

data ListTasks

See: newListTasks smart constructor.

Constructors

ListTasks' (Maybe TaskState) (Maybe Text) (Maybe Natural)

newListTasks :: ListTasks

Create a value of ListTasks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:ListTasks', listTasks_state - A structure used to filter the list of tasks.

$sel:nextToken:ListTasks', listTasks_nextToken - A pagination token to continue to the next page of tasks.

$sel:maxResults:ListTasks', listTasks_maxResults - The maximum number of tasks per page.

data ListTasksResponse

See: newListTasksResponse smart constructor.

Constructors

ListTasksResponse' (Maybe [TaskSummary]) (Maybe Text) Int

newListTasksResponse

Arguments

:: Int
$sel:httpStatus:ListTasksResponse'

-> ListTasksResponse

Create a value of ListTasksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tasks:ListTasksResponse', listTasksResponse_tasks - A list of task structures containing details about each task.

$sel:nextToken:ListTasks', listTasksResponse_nextToken - A pagination token to continue to the next page of tasks.

$sel:httpStatus:ListTasksResponse', listTasksResponse_httpStatus - The response's http status code.

DescribeExecution

data DescribeExecution

See: newDescribeExecution smart constructor.

Constructors

DescribeExecution' Text Text

newDescribeExecution

Arguments

:: Text
$sel:managedDeviceId:DescribeExecution'

-> Text
$sel:taskId:DescribeExecution'

-> DescribeExecution

Create a value of DescribeExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:managedDeviceId:DescribeExecution', describeExecution_managedDeviceId - The ID of the managed device.

$sel:taskId:DescribeExecution', describeExecution_taskId - The ID of the task that the action is describing.

data DescribeExecutionResponse

See: newDescribeExecutionResponse smart constructor.

Constructors

DescribeExecutionResponse' (Maybe Text) (Maybe ExecutionState) (Maybe POSIX) (Maybe Text) (Maybe POSIX) (Maybe Text) Int

newDescribeExecutionResponse

Arguments

:: Int
$sel:httpStatus:DescribeExecutionResponse'

-> DescribeExecutionResponse

Create a value of DescribeExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:executionId:DescribeExecutionResponse', describeExecutionResponse_executionId - The ID of the execution.

$sel:state:DescribeExecutionResponse', describeExecutionResponse_state - The current state of the execution.

$sel:lastUpdatedAt:DescribeExecutionResponse', describeExecutionResponse_lastUpdatedAt - When the status of the execution was last updated.

$sel:taskId:DescribeExecution', describeExecutionResponse_taskId - The ID of the task being executed on the device.

$sel:startedAt:DescribeExecutionResponse', describeExecutionResponse_startedAt - When the execution began.

$sel:managedDeviceId:DescribeExecution', describeExecutionResponse_managedDeviceId - The ID of the managed device that the task is being executed on.

$sel:httpStatus:DescribeExecutionResponse', describeExecutionResponse_httpStatus - The response's http status code.

DescribeDevice

data DescribeDevice

See: newDescribeDevice smart constructor.

Constructors

DescribeDevice' Text

newDescribeDevice

Arguments

:: Text
$sel:managedDeviceId:DescribeDevice'

-> DescribeDevice

Create a value of DescribeDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:managedDeviceId:DescribeDevice', describeDevice_managedDeviceId - The ID of the device that you are checking the information of.

data DescribeDeviceResponse

See: newDescribeDeviceResponse smart constructor.

Constructors

DescribeDeviceResponse' (Maybe UnlockState) (Maybe [Capacity]) (Maybe POSIX) (Maybe Text) (Maybe POSIX) (Maybe SoftwareInformation) (Maybe [PhysicalNetworkInterface]) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) Int

newDescribeDeviceResponse

Arguments

:: Int
$sel:httpStatus:DescribeDeviceResponse'

-> DescribeDeviceResponse

Create a value of DescribeDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceState:DescribeDeviceResponse', describeDeviceResponse_deviceState - The current state of the device.

$sel:deviceCapacities:DescribeDeviceResponse', describeDeviceResponse_deviceCapacities - The hardware specifications of the device.

$sel:lastUpdatedAt:DescribeDeviceResponse', describeDeviceResponse_lastUpdatedAt - When the device last pushed an update to the Amazon Web Services Cloud. Indicates when the device cache was refreshed.

$sel:associatedWithJob:DescribeDeviceResponse', describeDeviceResponse_associatedWithJob - The ID of the job used when ordering the device.

$sel:lastReachedOutAt:DescribeDeviceResponse', describeDeviceResponse_lastReachedOutAt - When the device last contacted the Amazon Web Services Cloud. Indicates that the device is online.

$sel:software:DescribeDeviceResponse', describeDeviceResponse_software - The software installed on the device.

$sel:physicalNetworkInterfaces:DescribeDeviceResponse', describeDeviceResponse_physicalNetworkInterfaces - The network interfaces available on the device.

$sel:managedDeviceId:DescribeDevice', describeDeviceResponse_managedDeviceId - The ID of the device that you checked the information for.

$sel:managedDeviceArn:DescribeDeviceResponse', describeDeviceResponse_managedDeviceArn - The Amazon Resource Name (ARN) of the device.

$sel:deviceType:DescribeDeviceResponse', describeDeviceResponse_deviceType - The type of Amazon Web Services Snow Family device.

$sel:tags:DescribeDeviceResponse', describeDeviceResponse_tags - Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

$sel:httpStatus:DescribeDeviceResponse', describeDeviceResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The Amazon Resource Name (ARN) of the device or task.

$sel:tags:TagResource', tagResource_tags - Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse'

newTagResourceResponse :: TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

CancelTask

data CancelTask

See: newCancelTask smart constructor.

Constructors

CancelTask' Text

newCancelTask

Arguments

:: Text
$sel:taskId:CancelTask'

-> CancelTask

Create a value of CancelTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskId:CancelTask', cancelTask_taskId - The ID of the task that you are attempting to cancel. You can retrieve a task ID by using the ListTasks operation.

data CancelTaskResponse

See: newCancelTaskResponse smart constructor.

Constructors

CancelTaskResponse' (Maybe Text) Int

newCancelTaskResponse

Arguments

:: Int
$sel:httpStatus:CancelTaskResponse'

-> CancelTaskResponse

Create a value of CancelTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskId:CancelTask', cancelTaskResponse_taskId - The ID of the task that you are attempting to cancel.

$sel:httpStatus:CancelTaskResponse', cancelTaskResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text [Text]

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - The Amazon Resource Name (ARN) of the device or task.

$sel:tagKeys:UntagResource', untagResource_tagKeys - Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse'

newUntagResourceResponse :: UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

ListDevices (Paginated)

data ListDevices

See: newListDevices smart constructor.

Constructors

ListDevices' (Maybe Text) (Maybe Text) (Maybe Natural)

newListDevices :: ListDevices

Create a value of ListDevices with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:ListDevices', listDevices_jobId - The ID of the job used to order the device.

$sel:nextToken:ListDevices', listDevices_nextToken - A pagination token to continue to the next page of results.

$sel:maxResults:ListDevices', listDevices_maxResults - The maximum number of devices to list per page.

data ListDevicesResponse

See: newListDevicesResponse smart constructor.

Constructors

ListDevicesResponse' (Maybe Text) (Maybe [DeviceSummary]) Int

newListDevicesResponse

Arguments

:: Int
$sel:httpStatus:ListDevicesResponse'

-> ListDevicesResponse

Create a value of ListDevicesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDevices', listDevicesResponse_nextToken - A pagination token to continue to the next page of devices.

$sel:devices:ListDevicesResponse', listDevicesResponse_devices - A list of device structures that contain information about the device.

$sel:httpStatus:ListDevicesResponse', listDevicesResponse_httpStatus - The response's http status code.

Types

AttachmentStatus

newtype AttachmentStatus

Constructors

AttachmentStatus'

Fields

fromAttachmentStatus :: Text

Bundled Patterns

pattern AttachmentStatus_ATTACHED :: AttachmentStatus

pattern AttachmentStatus_ATTACHING :: AttachmentStatus

pattern AttachmentStatus_DETACHED :: AttachmentStatus

pattern AttachmentStatus_DETACHING :: AttachmentStatus

ExecutionState

newtype ExecutionState

Constructors

ExecutionState'

Fields

fromExecutionState :: Text

Bundled Patterns

pattern ExecutionState_CANCELED :: ExecutionState

pattern ExecutionState_FAILED :: ExecutionState

pattern ExecutionState_IN_PROGRESS :: ExecutionState

pattern ExecutionState_QUEUED :: ExecutionState

pattern ExecutionState_REJECTED :: ExecutionState

pattern ExecutionState_SUCCEEDED :: ExecutionState

pattern ExecutionState_TIMED_OUT :: ExecutionState

InstanceStateName

newtype InstanceStateName

Constructors

InstanceStateName'

Fields

fromInstanceStateName :: Text

Bundled Patterns

pattern InstanceStateName_PENDING :: InstanceStateName

pattern InstanceStateName_RUNNING :: InstanceStateName

pattern InstanceStateName_SHUTTING_DOWN :: InstanceStateName

pattern InstanceStateName_STOPPED :: InstanceStateName

pattern InstanceStateName_STOPPING :: InstanceStateName

pattern InstanceStateName_TERMINATED :: InstanceStateName

IpAddressAssignment

newtype IpAddressAssignment

Constructors

IpAddressAssignment'

Fields

fromIpAddressAssignment :: Text

Bundled Patterns

pattern IpAddressAssignment_DHCP :: IpAddressAssignment

pattern IpAddressAssignment_STATIC :: IpAddressAssignment

PhysicalConnectorType

newtype PhysicalConnectorType

Constructors

PhysicalConnectorType'

Fields

fromPhysicalConnectorType :: Text

Bundled Patterns

pattern PhysicalConnectorType_QSFP :: PhysicalConnectorType

pattern PhysicalConnectorType_RJ45 :: PhysicalConnectorType

pattern PhysicalConnectorType_RJ45_2 :: PhysicalConnectorType

pattern PhysicalConnectorType_SFP_PLUS :: PhysicalConnectorType

pattern PhysicalConnectorType_WIFI :: PhysicalConnectorType

TaskState

newtype TaskState

Constructors

TaskState'

Fields

fromTaskState :: Text

Bundled Patterns

pattern TaskState_CANCELED :: TaskState

pattern TaskState_COMPLETED :: TaskState

pattern TaskState_IN_PROGRESS :: TaskState

UnlockState

newtype UnlockState

Constructors

UnlockState'

Fields

fromUnlockState :: Text

Bundled Patterns

pattern UnlockState_LOCKED :: UnlockState

pattern UnlockState_UNLOCKED :: UnlockState

pattern UnlockState_UNLOCKING :: UnlockState

Capacity

data Capacity

The physical capacity of the Amazon Web Services Snow Family device.

See: newCapacity smart constructor.

Constructors

Capacity' (Maybe Integer) (Maybe Text) (Maybe Integer) (Maybe Text) (Maybe Integer)

newCapacity :: Capacity

Create a value of Capacity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:used:Capacity', capacity_used - The amount of capacity used on the device.

$sel:name:Capacity', capacity_name - The name of the type of capacity, such as memory.

$sel:total:Capacity', capacity_total - The total capacity on the device.

$sel:unit:Capacity', capacity_unit - The unit of measure for the type of capacity.

$sel:available:Capacity', capacity_available - The amount of capacity available for use on the device.

Command

data Command

The command given to the device to execute.

See: newCommand smart constructor.

Constructors

Command' (Maybe Unlock) (Maybe Reboot)

newCommand :: Command

Create a value of Command with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unlock:Command', command_unlock - Unlocks the device.

$sel:reboot:Command', command_reboot - Reboots the device.

CpuOptions

data CpuOptions

The options for how a device's CPU is configured.

See: newCpuOptions smart constructor.

Constructors

CpuOptions' (Maybe Int) (Maybe Int)

newCpuOptions :: CpuOptions

Create a value of CpuOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:coreCount:CpuOptions', cpuOptions_coreCount - The number of cores that the CPU can use.

$sel:threadsPerCore:CpuOptions', cpuOptions_threadsPerCore - The number of threads per core in the CPU.

DeviceSummary

data DeviceSummary

Identifying information about the device.

See: newDeviceSummary smart constructor.

Constructors

DeviceSummary' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text))

newDeviceSummary :: DeviceSummary

Create a value of DeviceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associatedWithJob:DeviceSummary', deviceSummary_associatedWithJob - The ID of the job used to order the device.

$sel:managedDeviceId:DeviceSummary', deviceSummary_managedDeviceId - The ID of the device.

$sel:managedDeviceArn:DeviceSummary', deviceSummary_managedDeviceArn - The Amazon Resource Name (ARN) of the device.

$sel:tags:DeviceSummary', deviceSummary_tags - Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

EbsInstanceBlockDevice

data EbsInstanceBlockDevice

Describes a parameter used to set up an Amazon Elastic Block Store (Amazon EBS) volume in a block device mapping.

See: newEbsInstanceBlockDevice smart constructor.

Constructors

EbsInstanceBlockDevice' (Maybe Bool) (Maybe AttachmentStatus) (Maybe Text) (Maybe POSIX)

newEbsInstanceBlockDevice :: EbsInstanceBlockDevice

Create a value of EbsInstanceBlockDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deleteOnTermination:EbsInstanceBlockDevice', ebsInstanceBlockDevice_deleteOnTermination - A value that indicates whether the volume is deleted on instance termination.

$sel:status:EbsInstanceBlockDevice', ebsInstanceBlockDevice_status - The attachment state.

$sel:volumeId:EbsInstanceBlockDevice', ebsInstanceBlockDevice_volumeId - The ID of the Amazon EBS volume.

$sel:attachTime:EbsInstanceBlockDevice', ebsInstanceBlockDevice_attachTime - When the attachment was initiated.

ExecutionSummary

data ExecutionSummary

The summary of a task execution on a specified device.

See: newExecutionSummary smart constructor.

Constructors

ExecutionSummary' (Maybe Text) (Maybe ExecutionState) (Maybe Text) (Maybe Text)

newExecutionSummary :: ExecutionSummary

Create a value of ExecutionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:executionId:ExecutionSummary', executionSummary_executionId - The ID of the execution.

$sel:state:ExecutionSummary', executionSummary_state - The state of the execution.

$sel:taskId:ExecutionSummary', executionSummary_taskId - The ID of the task.

$sel:managedDeviceId:ExecutionSummary', executionSummary_managedDeviceId - The ID of the managed device that the task is being executed on.

Instance

data Instance

The description of an instance. Currently, Amazon EC2 instances are the only supported instance type.

See: newInstance smart constructor.

Constructors

Instance' (Maybe Text) (Maybe InstanceState) (Maybe [SecurityGroupIdentifier]) (Maybe POSIX) (Maybe CpuOptions) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe POSIX) (Maybe [InstanceBlockDeviceMapping]) (Maybe Text) (Maybe Int)

newInstance :: Instance

Create a value of Instance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:Instance', instance_instanceId - The ID of the instance.

$sel:state:Instance', instance_state - Undocumented member.

$sel:securityGroups:Instance', instance_securityGroups - The security groups for the instance.

$sel:createdAt:Instance', instance_createdAt - When the instance was created.

$sel:cpuOptions:Instance', instance_cpuOptions - The CPU options for the instance.

$sel:rootDeviceName:Instance', instance_rootDeviceName - The device name of the root device volume (for example, /dev/sda1).

$sel:instanceType:Instance', instance_instanceType - The instance type.

$sel:imageId:Instance', instance_imageId - The ID of the AMI used to launch the instance.

$sel:privateIpAddress:Instance', instance_privateIpAddress - The private IPv4 address assigned to the instance.

$sel:updatedAt:Instance', instance_updatedAt - When the instance was last updated.

$sel:blockDeviceMappings:Instance', instance_blockDeviceMappings - Any block device mapping entries for the instance.

$sel:publicIpAddress:Instance', instance_publicIpAddress - The public IPv4 address assigned to the instance.

$sel:amiLaunchIndex:Instance', instance_amiLaunchIndex - The Amazon Machine Image (AMI) launch index, which you can use to find this instance in the launch group.

InstanceBlockDeviceMapping

data InstanceBlockDeviceMapping

The description of a block device mapping.

See: newInstanceBlockDeviceMapping smart constructor.

Constructors

InstanceBlockDeviceMapping' (Maybe EbsInstanceBlockDevice) (Maybe Text)

newInstanceBlockDeviceMapping :: InstanceBlockDeviceMapping

Create a value of InstanceBlockDeviceMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ebs:InstanceBlockDeviceMapping', instanceBlockDeviceMapping_ebs - The parameters used to automatically set up Amazon Elastic Block Store (Amazon EBS) volumes when the instance is launched.

$sel:deviceName:InstanceBlockDeviceMapping', instanceBlockDeviceMapping_deviceName - The block device name.

InstanceState

data InstanceState

The description of the current state of an instance.

See: newInstanceState smart constructor.

Constructors

InstanceState' (Maybe InstanceStateName) (Maybe Int)

newInstanceState :: InstanceState

Create a value of InstanceState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:InstanceState', instanceState_name - The current state of the instance.

$sel:code:InstanceState', instanceState_code - The state of the instance as a 16-bit unsigned integer.

The high byte is all of the bits between 2^8 and (2^16)-1, which equals decimal values between 256 and 65,535. These numerical values are used for internal purposes and should be ignored.

The low byte is all of the bits between 2^0 and (2^8)-1, which equals decimal values between 0 and 255.

The valid values for the instance state code are all in the range of the low byte. These values are:

0 : pending

16 : running

32 : shutting-down

48 : terminated

64 : stopping

80 : stopped

You can ignore the high byte value by zeroing out all of the bits above 2^8 or 256 in decimal.

InstanceSummary

data InstanceSummary

The details about the instance.

See: newInstanceSummary smart constructor.

Constructors

InstanceSummary' (Maybe POSIX) (Maybe Instance)

newInstanceSummary :: InstanceSummary

Create a value of InstanceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedAt:InstanceSummary', instanceSummary_lastUpdatedAt - When the instance summary was last updated.

$sel:instance':InstanceSummary', instanceSummary_instance - A structure containing details about the instance.

PhysicalNetworkInterface

data PhysicalNetworkInterface

The details about the physical network interface for the device.

See: newPhysicalNetworkInterface smart constructor.

Constructors

PhysicalNetworkInterface' (Maybe Text) (Maybe Text) (Maybe IpAddressAssignment) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe PhysicalConnectorType)

newPhysicalNetworkInterface :: PhysicalNetworkInterface

Create a value of PhysicalNetworkInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipAddress:PhysicalNetworkInterface', physicalNetworkInterface_ipAddress - The IP address of the device.

$sel:macAddress:PhysicalNetworkInterface', physicalNetworkInterface_macAddress - The MAC address of the device.

$sel:ipAddressAssignment:PhysicalNetworkInterface', physicalNetworkInterface_ipAddressAssignment - A value that describes whether the IP address is dynamic or persistent.

$sel:defaultGateway:PhysicalNetworkInterface', physicalNetworkInterface_defaultGateway - The default gateway of the device.

$sel:physicalNetworkInterfaceId:PhysicalNetworkInterface', physicalNetworkInterface_physicalNetworkInterfaceId - The physical network interface ID.

$sel:netmask:PhysicalNetworkInterface', physicalNetworkInterface_netmask - The netmask used to divide the IP address into subnets.

$sel:physicalConnectorType:PhysicalNetworkInterface', physicalNetworkInterface_physicalConnectorType - The physical connector type.

Reboot

data Reboot

A structure used to reboot the device.

See: newReboot smart constructor.

Constructors

Reboot'

newReboot :: Reboot

Create a value of Reboot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

ResourceSummary

data ResourceSummary

A summary of a resource available on the device.

See: newResourceSummary smart constructor.

Constructors

ResourceSummary' (Maybe Text) (Maybe Text) Text

newResourceSummary

Arguments

:: Text
$sel:resourceType:ResourceSummary'

-> ResourceSummary

Create a value of ResourceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ResourceSummary', resourceSummary_arn - The Amazon Resource Name (ARN) of the resource.

$sel:id:ResourceSummary', resourceSummary_id - The ID of the resource.

$sel:resourceType:ResourceSummary', resourceSummary_resourceType - The resource type.

SecurityGroupIdentifier

data SecurityGroupIdentifier

Information about the device's security group.

See: newSecurityGroupIdentifier smart constructor.

Constructors

SecurityGroupIdentifier' (Maybe Text) (Maybe Text)

newSecurityGroupIdentifier :: SecurityGroupIdentifier

Create a value of SecurityGroupIdentifier with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupId:SecurityGroupIdentifier', securityGroupIdentifier_groupId - The security group ID.

$sel:groupName:SecurityGroupIdentifier', securityGroupIdentifier_groupName - The security group name.

SoftwareInformation

data SoftwareInformation

Information about the software on the device.

See: newSoftwareInformation smart constructor.

Constructors

SoftwareInformation' (Maybe Text) (Maybe Text) (Maybe Text)

newSoftwareInformation :: SoftwareInformation

Create a value of SoftwareInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:installedVersion:SoftwareInformation', softwareInformation_installedVersion - The version of the software currently installed on the device.

$sel:installState:SoftwareInformation', softwareInformation_installState - The state of the software that is installed or that is being installed on the device.

$sel:installingVersion:SoftwareInformation', softwareInformation_installingVersion - The version of the software being installed on the device.

TaskSummary

data TaskSummary

Information about the task assigned to one or many devices.

See: newTaskSummary smart constructor.

Constructors

TaskSummary' (Maybe TaskState) (Maybe Text) (Maybe (HashMap Text Text)) Text

newTaskSummary

Arguments

:: Text
$sel:taskId:TaskSummary'

-> TaskSummary

Create a value of TaskSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:TaskSummary', taskSummary_state - The state of the task assigned to one or many devices.

$sel:taskArn:TaskSummary', taskSummary_taskArn - The Amazon Resource Name (ARN) of the task.

$sel:tags:TaskSummary', taskSummary_tags - Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

$sel:taskId:TaskSummary', taskSummary_taskId - The task ID.

Unlock

data Unlock

A structure used to unlock a device.

See: newUnlock smart constructor.

Constructors

Unlock'

newUnlock :: Unlock

Create a value of Unlock with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.