Network.AWS.SnowDeviceManagement.CreateTask

Description

Instructs one or more devices to start a task, such as unlocking or rebooting.

Creating a Request

data CreateTask

See: newCreateTask smart constructor.

Constructors

CreateTask'

Fields

clientToken :: Maybe Text

A token ensuring that the action is called only once with the specified details.

description :: Maybe Text

A description of the task and its targets.

tags :: Maybe (HashMap Text Text)

Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

command :: Command

The task to be performed. Only one task is executed on a device at a time.

targets :: NonEmpty Text

A list of managed device IDs.

newCreateTask

Arguments

:: Command
$sel:command:CreateTask'

-> NonEmpty Text
$sel:targets:CreateTask'

-> CreateTask

Create a value of CreateTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateTask', createTask_clientToken - A token ensuring that the action is called only once with the specified details.

$sel:description:CreateTask', createTask_description - A description of the task and its targets.

$sel:tags:CreateTask', createTask_tags - Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

$sel:command:CreateTask', createTask_command - The task to be performed. Only one task is executed on a device at a time.

$sel:targets:CreateTask', createTask_targets - A list of managed device IDs.

Request Lenses

createTask_clientToken :: Lens' CreateTask (Maybe Text)

A token ensuring that the action is called only once with the specified details.

createTask_description :: Lens' CreateTask (Maybe Text)

A description of the task and its targets.

createTask_tags :: Lens' CreateTask (Maybe (HashMap Text Text))

Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

createTask_command :: Lens' CreateTask Command

The task to be performed. Only one task is executed on a device at a time.

createTask_targets :: Lens' CreateTask (NonEmpty Text)

A list of managed device IDs.

Destructuring the Response

data CreateTaskResponse

See: newCreateTaskResponse smart constructor.

Constructors

CreateTaskResponse'

Fields

taskId :: Maybe Text

The ID of the task that you created.

taskArn :: Maybe Text

The Amazon Resource Name (ARN) of the task that you created.

httpStatus :: Int

The response's http status code.

newCreateTaskResponse

Arguments

:: Int
$sel:httpStatus:CreateTaskResponse'

-> CreateTaskResponse

Create a value of CreateTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskId:CreateTaskResponse', createTaskResponse_taskId - The ID of the task that you created.

$sel:taskArn:CreateTaskResponse', createTaskResponse_taskArn - The Amazon Resource Name (ARN) of the task that you created.

$sel:httpStatus:CreateTaskResponse', createTaskResponse_httpStatus - The response's http status code.

Response Lenses

createTaskResponse_taskId :: Lens' CreateTaskResponse (Maybe Text)

The ID of the task that you created.

createTaskResponse_taskArn :: Lens' CreateTaskResponse (Maybe Text)

The Amazon Resource Name (ARN) of the task that you created.

createTaskResponse_httpStatus :: Lens' CreateTaskResponse Int

The response's http status code.