Network.AWS.SnowDeviceManagement.DescribeDeviceEc2Instances

Description

Checks the current state of the Amazon EC2 instances. The output is similar to describeDevice, but the results are sourced from the device cache in the Amazon Web Services Cloud and include a subset of the available fields.

Creating a Request

data DescribeDeviceEc2Instances

See: newDescribeDeviceEc2Instances smart constructor.

Constructors

DescribeDeviceEc2Instances'

Fields

instanceIds :: [Text]

A list of instance IDs associated with the managed device.

managedDeviceId :: Text

The ID of the managed device.

newDescribeDeviceEc2Instances

Arguments

:: Text
$sel:managedDeviceId:DescribeDeviceEc2Instances'

-> DescribeDeviceEc2Instances

Create a value of DescribeDeviceEc2Instances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceIds:DescribeDeviceEc2Instances', describeDeviceEc2Instances_instanceIds - A list of instance IDs associated with the managed device.

$sel:managedDeviceId:DescribeDeviceEc2Instances', describeDeviceEc2Instances_managedDeviceId - The ID of the managed device.

Request Lenses

describeDeviceEc2Instances_instanceIds :: Lens' DescribeDeviceEc2Instances [Text]

A list of instance IDs associated with the managed device.

describeDeviceEc2Instances_managedDeviceId :: Lens' DescribeDeviceEc2Instances Text

The ID of the managed device.

Destructuring the Response

data DescribeDeviceEc2InstancesResponse

See: newDescribeDeviceEc2InstancesResponse smart constructor.

Constructors

DescribeDeviceEc2InstancesResponse'

Fields

instances :: Maybe [InstanceSummary]

A list of structures containing information about each instance.

httpStatus :: Int

The response's http status code.

newDescribeDeviceEc2InstancesResponse

Arguments

:: Int
$sel:httpStatus:DescribeDeviceEc2InstancesResponse'

-> DescribeDeviceEc2InstancesResponse

Create a value of DescribeDeviceEc2InstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instances:DescribeDeviceEc2InstancesResponse', describeDeviceEc2InstancesResponse_instances - A list of structures containing information about each instance.

$sel:httpStatus:DescribeDeviceEc2InstancesResponse', describeDeviceEc2InstancesResponse_httpStatus - The response's http status code.

Response Lenses

describeDeviceEc2InstancesResponse_instances :: Lens' DescribeDeviceEc2InstancesResponse (Maybe [InstanceSummary])

A list of structures containing information about each instance.

describeDeviceEc2InstancesResponse_httpStatus :: Lens' DescribeDeviceEc2InstancesResponse Int

The response's http status code.