Network.AWS.SnowDeviceManagement.DescribeTask

Description

Checks the metadata for a given task on a device.

Creating a Request

data DescribeTask

See: newDescribeTask smart constructor.

Constructors

DescribeTask'

Fields

taskId :: Text

The ID of the task to be described.

newDescribeTask

Arguments

:: Text
$sel:taskId:DescribeTask'

-> DescribeTask

Create a value of DescribeTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskId:DescribeTask', describeTask_taskId - The ID of the task to be described.

Request Lenses

describeTask_taskId :: Lens' DescribeTask Text

The ID of the task to be described.

Destructuring the Response

data DescribeTaskResponse

See: newDescribeTaskResponse smart constructor.

Constructors

DescribeTaskResponse'

Fields

state :: Maybe TaskState

The current state of the task.

lastUpdatedAt :: Maybe POSIX

When the state of the task was last updated.

createdAt :: Maybe POSIX

When the CreateTask operation was called.

taskId :: Maybe Text

The ID of the task.

taskArn :: Maybe Text

The Amazon Resource Name (ARN) of the task.

targets :: Maybe (NonEmpty Text)

The managed devices that the task was sent to.

completedAt :: Maybe POSIX

When the task was completed.

description :: Maybe Text

The description provided of the task and managed devices.

tags :: Maybe (HashMap Text Text)

Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

httpStatus :: Int

The response's http status code.

newDescribeTaskResponse

Arguments

:: Int
$sel:httpStatus:DescribeTaskResponse'

-> DescribeTaskResponse

Create a value of DescribeTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:DescribeTaskResponse', describeTaskResponse_state - The current state of the task.

$sel:lastUpdatedAt:DescribeTaskResponse', describeTaskResponse_lastUpdatedAt - When the state of the task was last updated.

$sel:createdAt:DescribeTaskResponse', describeTaskResponse_createdAt - When the CreateTask operation was called.

$sel:taskId:DescribeTask', describeTaskResponse_taskId - The ID of the task.

$sel:taskArn:DescribeTaskResponse', describeTaskResponse_taskArn - The Amazon Resource Name (ARN) of the task.

$sel:targets:DescribeTaskResponse', describeTaskResponse_targets - The managed devices that the task was sent to.

$sel:completedAt:DescribeTaskResponse', describeTaskResponse_completedAt - When the task was completed.

$sel:description:DescribeTaskResponse', describeTaskResponse_description - The description provided of the task and managed devices.

$sel:tags:DescribeTaskResponse', describeTaskResponse_tags - Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

$sel:httpStatus:DescribeTaskResponse', describeTaskResponse_httpStatus - The response's http status code.

Response Lenses

describeTaskResponse_state :: Lens' DescribeTaskResponse (Maybe TaskState)

The current state of the task.

describeTaskResponse_lastUpdatedAt :: Lens' DescribeTaskResponse (Maybe UTCTime)

When the state of the task was last updated.

describeTaskResponse_createdAt :: Lens' DescribeTaskResponse (Maybe UTCTime)

When the CreateTask operation was called.

describeTaskResponse_taskId :: Lens' DescribeTaskResponse (Maybe Text)

The ID of the task.

describeTaskResponse_taskArn :: Lens' DescribeTaskResponse (Maybe Text)

The Amazon Resource Name (ARN) of the task.

describeTaskResponse_targets :: Lens' DescribeTaskResponse (Maybe (NonEmpty Text))

The managed devices that the task was sent to.

describeTaskResponse_completedAt :: Lens' DescribeTaskResponse (Maybe UTCTime)

When the task was completed.

describeTaskResponse_description :: Lens' DescribeTaskResponse (Maybe Text)

The description provided of the task and managed devices.

describeTaskResponse_tags :: Lens' DescribeTaskResponse (Maybe (HashMap Text Text))

Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

describeTaskResponse_httpStatus :: Lens' DescribeTaskResponse Int

The response's http status code.