Network.AWS.SnowDeviceManagement.Lens

Description

Operations

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the device or task.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The list of tags for the device or task.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

DescribeTask

describeTask_taskId :: Lens' DescribeTask Text

The ID of the task to be described.

describeTaskResponse_state :: Lens' DescribeTaskResponse (Maybe TaskState)

The current state of the task.

describeTaskResponse_lastUpdatedAt :: Lens' DescribeTaskResponse (Maybe UTCTime)

When the state of the task was last updated.

describeTaskResponse_createdAt :: Lens' DescribeTaskResponse (Maybe UTCTime)

When the CreateTask operation was called.

describeTaskResponse_taskId :: Lens' DescribeTaskResponse (Maybe Text)

The ID of the task.

describeTaskResponse_taskArn :: Lens' DescribeTaskResponse (Maybe Text)

The Amazon Resource Name (ARN) of the task.

describeTaskResponse_targets :: Lens' DescribeTaskResponse (Maybe (NonEmpty Text))

The managed devices that the task was sent to.

describeTaskResponse_completedAt :: Lens' DescribeTaskResponse (Maybe UTCTime)

When the task was completed.

describeTaskResponse_description :: Lens' DescribeTaskResponse (Maybe Text)

The description provided of the task and managed devices.

describeTaskResponse_tags :: Lens' DescribeTaskResponse (Maybe (HashMap Text Text))

Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

describeTaskResponse_httpStatus :: Lens' DescribeTaskResponse Int

The response's http status code.

ListDeviceResources

listDeviceResources_nextToken :: Lens' ListDeviceResources (Maybe Text)

A pagination token to continue to the next page of results.

listDeviceResources_type :: Lens' ListDeviceResources (Maybe Text)

A structure used to filter the results by type of resource.

listDeviceResources_maxResults :: Lens' ListDeviceResources (Maybe Natural)

The maximum number of resources per page.

listDeviceResources_managedDeviceId :: Lens' ListDeviceResources Text

The ID of the managed device that you are listing the resources of.

listDeviceResourcesResponse_resources :: Lens' ListDeviceResourcesResponse (Maybe [ResourceSummary])

A structure defining the resource's type, Amazon Resource Name (ARN), and ID.

listDeviceResourcesResponse_nextToken :: Lens' ListDeviceResourcesResponse (Maybe Text)

A pagination token to continue to the next page of results.

listDeviceResourcesResponse_httpStatus :: Lens' ListDeviceResourcesResponse Int

The response's http status code.

ListExecutions

listExecutions_state :: Lens' ListExecutions (Maybe ExecutionState)

A structure used to filter the tasks by their current state.

listExecutions_nextToken :: Lens' ListExecutions (Maybe Text)

A pagination token to continue to the next page of tasks.

listExecutions_maxResults :: Lens' ListExecutions (Maybe Natural)

The maximum number of tasks to list per page.

listExecutions_taskId :: Lens' ListExecutions Text

The ID of the task.

listExecutionsResponse_executions :: Lens' ListExecutionsResponse (Maybe [ExecutionSummary])

A list of executions. Each execution contains the task ID, the device that the task is executing on, the execution ID, and the status of the execution.

listExecutionsResponse_nextToken :: Lens' ListExecutionsResponse (Maybe Text)

A pagination token to continue to the next page of executions.

listExecutionsResponse_httpStatus :: Lens' ListExecutionsResponse Int

The response's http status code.

DescribeDeviceEc2Instances

describeDeviceEc2Instances_instanceIds :: Lens' DescribeDeviceEc2Instances [Text]

A list of instance IDs associated with the managed device.

describeDeviceEc2Instances_managedDeviceId :: Lens' DescribeDeviceEc2Instances Text

The ID of the managed device.

describeDeviceEc2InstancesResponse_instances :: Lens' DescribeDeviceEc2InstancesResponse (Maybe [InstanceSummary])

A list of structures containing information about each instance.

describeDeviceEc2InstancesResponse_httpStatus :: Lens' DescribeDeviceEc2InstancesResponse Int

The response's http status code.

CreateTask

createTask_clientToken :: Lens' CreateTask (Maybe Text)

A token ensuring that the action is called only once with the specified details.

createTask_description :: Lens' CreateTask (Maybe Text)

A description of the task and its targets.

createTask_tags :: Lens' CreateTask (Maybe (HashMap Text Text))

Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

createTask_command :: Lens' CreateTask Command

The task to be performed. Only one task is executed on a device at a time.

createTask_targets :: Lens' CreateTask (NonEmpty Text)

A list of managed device IDs.

createTaskResponse_taskId :: Lens' CreateTaskResponse (Maybe Text)

The ID of the task that you created.

createTaskResponse_taskArn :: Lens' CreateTaskResponse (Maybe Text)

The Amazon Resource Name (ARN) of the task that you created.

createTaskResponse_httpStatus :: Lens' CreateTaskResponse Int

The response's http status code.

ListTasks

listTasks_state :: Lens' ListTasks (Maybe TaskState)

A structure used to filter the list of tasks.

listTasks_nextToken :: Lens' ListTasks (Maybe Text)

A pagination token to continue to the next page of tasks.

listTasks_maxResults :: Lens' ListTasks (Maybe Natural)

The maximum number of tasks per page.

listTasksResponse_tasks :: Lens' ListTasksResponse (Maybe [TaskSummary])

A list of task structures containing details about each task.

listTasksResponse_nextToken :: Lens' ListTasksResponse (Maybe Text)

A pagination token to continue to the next page of tasks.

listTasksResponse_httpStatus :: Lens' ListTasksResponse Int

The response's http status code.

DescribeExecution

describeExecution_managedDeviceId :: Lens' DescribeExecution Text

The ID of the managed device.

describeExecution_taskId :: Lens' DescribeExecution Text

The ID of the task that the action is describing.

describeExecutionResponse_executionId :: Lens' DescribeExecutionResponse (Maybe Text)

The ID of the execution.

describeExecutionResponse_state :: Lens' DescribeExecutionResponse (Maybe ExecutionState)

The current state of the execution.

describeExecutionResponse_lastUpdatedAt :: Lens' DescribeExecutionResponse (Maybe UTCTime)

When the status of the execution was last updated.

describeExecutionResponse_taskId :: Lens' DescribeExecutionResponse (Maybe Text)

The ID of the task being executed on the device.

describeExecutionResponse_startedAt :: Lens' DescribeExecutionResponse (Maybe UTCTime)

When the execution began.

describeExecutionResponse_managedDeviceId :: Lens' DescribeExecutionResponse (Maybe Text)

The ID of the managed device that the task is being executed on.

describeExecutionResponse_httpStatus :: Lens' DescribeExecutionResponse Int

The response's http status code.

DescribeDevice

describeDevice_managedDeviceId :: Lens' DescribeDevice Text

The ID of the device that you are checking the information of.

describeDeviceResponse_deviceState :: Lens' DescribeDeviceResponse (Maybe UnlockState)

The current state of the device.

describeDeviceResponse_deviceCapacities :: Lens' DescribeDeviceResponse (Maybe [Capacity])

The hardware specifications of the device.

describeDeviceResponse_lastUpdatedAt :: Lens' DescribeDeviceResponse (Maybe UTCTime)

When the device last pushed an update to the Amazon Web Services Cloud. Indicates when the device cache was refreshed.

describeDeviceResponse_associatedWithJob :: Lens' DescribeDeviceResponse (Maybe Text)

The ID of the job used when ordering the device.

describeDeviceResponse_lastReachedOutAt :: Lens' DescribeDeviceResponse (Maybe UTCTime)

When the device last contacted the Amazon Web Services Cloud. Indicates that the device is online.

describeDeviceResponse_software :: Lens' DescribeDeviceResponse (Maybe SoftwareInformation)

The software installed on the device.

describeDeviceResponse_physicalNetworkInterfaces :: Lens' DescribeDeviceResponse (Maybe [PhysicalNetworkInterface])

The network interfaces available on the device.

describeDeviceResponse_managedDeviceId :: Lens' DescribeDeviceResponse (Maybe Text)

The ID of the device that you checked the information for.

describeDeviceResponse_managedDeviceArn :: Lens' DescribeDeviceResponse (Maybe Text)

The Amazon Resource Name (ARN) of the device.

describeDeviceResponse_deviceType :: Lens' DescribeDeviceResponse (Maybe Text)

The type of Amazon Web Services Snow Family device.

describeDeviceResponse_tags :: Lens' DescribeDeviceResponse (Maybe (HashMap Text Text))

Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

describeDeviceResponse_httpStatus :: Lens' DescribeDeviceResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the device or task.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

CancelTask

cancelTask_taskId :: Lens' CancelTask Text

The ID of the task that you are attempting to cancel. You can retrieve a task ID by using the ListTasks operation.

cancelTaskResponse_taskId :: Lens' CancelTaskResponse (Maybe Text)

The ID of the task that you are attempting to cancel.

cancelTaskResponse_httpStatus :: Lens' CancelTaskResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the device or task.

untagResource_tagKeys :: Lens' UntagResource [Text]

Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

ListDevices

listDevices_jobId :: Lens' ListDevices (Maybe Text)

The ID of the job used to order the device.

listDevices_nextToken :: Lens' ListDevices (Maybe Text)

A pagination token to continue to the next page of results.

listDevices_maxResults :: Lens' ListDevices (Maybe Natural)

The maximum number of devices to list per page.

listDevicesResponse_nextToken :: Lens' ListDevicesResponse (Maybe Text)

A pagination token to continue to the next page of devices.

listDevicesResponse_devices :: Lens' ListDevicesResponse (Maybe [DeviceSummary])

A list of device structures that contain information about the device.

listDevicesResponse_httpStatus :: Lens' ListDevicesResponse Int

The response's http status code.

Types

Capacity

capacity_used :: Lens' Capacity (Maybe Integer)

The amount of capacity used on the device.

capacity_name :: Lens' Capacity (Maybe Text)

The name of the type of capacity, such as memory.

capacity_total :: Lens' Capacity (Maybe Integer)

The total capacity on the device.

capacity_unit :: Lens' Capacity (Maybe Text)

The unit of measure for the type of capacity.

capacity_available :: Lens' Capacity (Maybe Integer)

The amount of capacity available for use on the device.

Command

command_unlock :: Lens' Command (Maybe Unlock)

Unlocks the device.

command_reboot :: Lens' Command (Maybe Reboot)

Reboots the device.

CpuOptions

cpuOptions_coreCount :: Lens' CpuOptions (Maybe Int)

The number of cores that the CPU can use.

cpuOptions_threadsPerCore :: Lens' CpuOptions (Maybe Int)

The number of threads per core in the CPU.

DeviceSummary

deviceSummary_associatedWithJob :: Lens' DeviceSummary (Maybe Text)

The ID of the job used to order the device.

deviceSummary_managedDeviceId :: Lens' DeviceSummary (Maybe Text)

The ID of the device.

deviceSummary_managedDeviceArn :: Lens' DeviceSummary (Maybe Text)

The Amazon Resource Name (ARN) of the device.

deviceSummary_tags :: Lens' DeviceSummary (Maybe (HashMap Text Text))

Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

EbsInstanceBlockDevice

ebsInstanceBlockDevice_deleteOnTermination :: Lens' EbsInstanceBlockDevice (Maybe Bool)

A value that indicates whether the volume is deleted on instance termination.

ebsInstanceBlockDevice_status :: Lens' EbsInstanceBlockDevice (Maybe AttachmentStatus)

The attachment state.

ebsInstanceBlockDevice_volumeId :: Lens' EbsInstanceBlockDevice (Maybe Text)

The ID of the Amazon EBS volume.

ebsInstanceBlockDevice_attachTime :: Lens' EbsInstanceBlockDevice (Maybe UTCTime)

When the attachment was initiated.

ExecutionSummary

executionSummary_executionId :: Lens' ExecutionSummary (Maybe Text)

The ID of the execution.

executionSummary_state :: Lens' ExecutionSummary (Maybe ExecutionState)

The state of the execution.

executionSummary_taskId :: Lens' ExecutionSummary (Maybe Text)

The ID of the task.

executionSummary_managedDeviceId :: Lens' ExecutionSummary (Maybe Text)

The ID of the managed device that the task is being executed on.

Instance

instance_instanceId :: Lens' Instance (Maybe Text)

The ID of the instance.

instance_state :: Lens' Instance (Maybe InstanceState)

Undocumented member.

instance_securityGroups :: Lens' Instance (Maybe [SecurityGroupIdentifier])

The security groups for the instance.

instance_createdAt :: Lens' Instance (Maybe UTCTime)

When the instance was created.

instance_cpuOptions :: Lens' Instance (Maybe CpuOptions)

The CPU options for the instance.

instance_rootDeviceName :: Lens' Instance (Maybe Text)

The device name of the root device volume (for example, /dev/sda1).

instance_instanceType :: Lens' Instance (Maybe Text)

The instance type.

instance_imageId :: Lens' Instance (Maybe Text)

The ID of the AMI used to launch the instance.

instance_privateIpAddress :: Lens' Instance (Maybe Text)

The private IPv4 address assigned to the instance.

instance_updatedAt :: Lens' Instance (Maybe UTCTime)

When the instance was last updated.

instance_blockDeviceMappings :: Lens' Instance (Maybe [InstanceBlockDeviceMapping])

Any block device mapping entries for the instance.

instance_publicIpAddress :: Lens' Instance (Maybe Text)

The public IPv4 address assigned to the instance.

instance_amiLaunchIndex :: Lens' Instance (Maybe Int)

The Amazon Machine Image (AMI) launch index, which you can use to find this instance in the launch group.

InstanceBlockDeviceMapping

instanceBlockDeviceMapping_ebs :: Lens' InstanceBlockDeviceMapping (Maybe EbsInstanceBlockDevice)

The parameters used to automatically set up Amazon Elastic Block Store (Amazon EBS) volumes when the instance is launched.

instanceBlockDeviceMapping_deviceName :: Lens' InstanceBlockDeviceMapping (Maybe Text)

The block device name.

InstanceState

instanceState_name :: Lens' InstanceState (Maybe InstanceStateName)

The current state of the instance.

instanceState_code :: Lens' InstanceState (Maybe Int)

The state of the instance as a 16-bit unsigned integer.

The high byte is all of the bits between 2^8 and (2^16)-1, which equals decimal values between 256 and 65,535. These numerical values are used for internal purposes and should be ignored.

The low byte is all of the bits between 2^0 and (2^8)-1, which equals decimal values between 0 and 255.

The valid values for the instance state code are all in the range of the low byte. These values are:

0 : pending

16 : running

32 : shutting-down

48 : terminated

64 : stopping

80 : stopped

You can ignore the high byte value by zeroing out all of the bits above 2^8 or 256 in decimal.

InstanceSummary

instanceSummary_lastUpdatedAt :: Lens' InstanceSummary (Maybe UTCTime)

When the instance summary was last updated.

instanceSummary_instance :: Lens' InstanceSummary (Maybe Instance)

A structure containing details about the instance.

PhysicalNetworkInterface

physicalNetworkInterface_ipAddress :: Lens' PhysicalNetworkInterface (Maybe Text)

The IP address of the device.

physicalNetworkInterface_macAddress :: Lens' PhysicalNetworkInterface (Maybe Text)

The MAC address of the device.

physicalNetworkInterface_ipAddressAssignment :: Lens' PhysicalNetworkInterface (Maybe IpAddressAssignment)

A value that describes whether the IP address is dynamic or persistent.

physicalNetworkInterface_defaultGateway :: Lens' PhysicalNetworkInterface (Maybe Text)

The default gateway of the device.

physicalNetworkInterface_physicalNetworkInterfaceId :: Lens' PhysicalNetworkInterface (Maybe Text)

The physical network interface ID.

physicalNetworkInterface_netmask :: Lens' PhysicalNetworkInterface (Maybe Text)

The netmask used to divide the IP address into subnets.

physicalNetworkInterface_physicalConnectorType :: Lens' PhysicalNetworkInterface (Maybe PhysicalConnectorType)

The physical connector type.

Reboot

ResourceSummary

resourceSummary_arn :: Lens' ResourceSummary (Maybe Text)

The Amazon Resource Name (ARN) of the resource.

resourceSummary_id :: Lens' ResourceSummary (Maybe Text)

The ID of the resource.

resourceSummary_resourceType :: Lens' ResourceSummary Text

The resource type.

SecurityGroupIdentifier

securityGroupIdentifier_groupId :: Lens' SecurityGroupIdentifier (Maybe Text)

The security group ID.

securityGroupIdentifier_groupName :: Lens' SecurityGroupIdentifier (Maybe Text)

The security group name.

SoftwareInformation

softwareInformation_installedVersion :: Lens' SoftwareInformation (Maybe Text)

The version of the software currently installed on the device.

softwareInformation_installState :: Lens' SoftwareInformation (Maybe Text)

The state of the software that is installed or that is being installed on the device.

softwareInformation_installingVersion :: Lens' SoftwareInformation (Maybe Text)

The version of the software being installed on the device.

TaskSummary

taskSummary_state :: Lens' TaskSummary (Maybe TaskState)

The state of the task assigned to one or many devices.

taskSummary_taskArn :: Lens' TaskSummary (Maybe Text)

The Amazon Resource Name (ARN) of the task.

taskSummary_tags :: Lens' TaskSummary (Maybe (HashMap Text Text))

Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

taskSummary_taskId :: Lens' TaskSummary Text

The task ID.

Unlock