Network.AWS.SnowDeviceManagement.ListDeviceResources

Description

Returns a list of the Amazon Web Services resources available for a device. Currently, Amazon EC2 instances are the only supported resource type.

This operation returns paginated results.

Creating a Request

data ListDeviceResources

See: newListDeviceResources smart constructor.

Constructors

ListDeviceResources'

Fields

nextToken :: Maybe Text

A pagination token to continue to the next page of results.

type' :: Maybe Text

A structure used to filter the results by type of resource.

maxResults :: Maybe Natural

The maximum number of resources per page.

managedDeviceId :: Text

The ID of the managed device that you are listing the resources of.

newListDeviceResources

Arguments

:: Text
$sel:managedDeviceId:ListDeviceResources'

-> ListDeviceResources

Create a value of ListDeviceResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDeviceResources', listDeviceResources_nextToken - A pagination token to continue to the next page of results.

$sel:type':ListDeviceResources', listDeviceResources_type - A structure used to filter the results by type of resource.

$sel:maxResults:ListDeviceResources', listDeviceResources_maxResults - The maximum number of resources per page.

$sel:managedDeviceId:ListDeviceResources', listDeviceResources_managedDeviceId - The ID of the managed device that you are listing the resources of.

Request Lenses

listDeviceResources_nextToken :: Lens' ListDeviceResources (Maybe Text)

A pagination token to continue to the next page of results.

listDeviceResources_type :: Lens' ListDeviceResources (Maybe Text)

A structure used to filter the results by type of resource.

listDeviceResources_maxResults :: Lens' ListDeviceResources (Maybe Natural)

The maximum number of resources per page.

listDeviceResources_managedDeviceId :: Lens' ListDeviceResources Text

The ID of the managed device that you are listing the resources of.

Destructuring the Response

data ListDeviceResourcesResponse

See: newListDeviceResourcesResponse smart constructor.

Constructors

ListDeviceResourcesResponse'

Fields

resources :: Maybe [ResourceSummary]

A structure defining the resource's type, Amazon Resource Name (ARN), and ID.

nextToken :: Maybe Text

A pagination token to continue to the next page of results.

httpStatus :: Int

The response's http status code.

newListDeviceResourcesResponse

Arguments

:: Int
$sel:httpStatus:ListDeviceResourcesResponse'

-> ListDeviceResourcesResponse

Create a value of ListDeviceResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resources:ListDeviceResourcesResponse', listDeviceResourcesResponse_resources - A structure defining the resource's type, Amazon Resource Name (ARN), and ID.

$sel:nextToken:ListDeviceResources', listDeviceResourcesResponse_nextToken - A pagination token to continue to the next page of results.

$sel:httpStatus:ListDeviceResourcesResponse', listDeviceResourcesResponse_httpStatus - The response's http status code.

Response Lenses

listDeviceResourcesResponse_resources :: Lens' ListDeviceResourcesResponse (Maybe [ResourceSummary])

A structure defining the resource's type, Amazon Resource Name (ARN), and ID.

listDeviceResourcesResponse_nextToken :: Lens' ListDeviceResourcesResponse (Maybe Text)

A pagination token to continue to the next page of results.

listDeviceResourcesResponse_httpStatus :: Lens' ListDeviceResourcesResponse Int

The response's http status code.