Network.AWS.SnowDeviceManagement.Types

Description

Service Configuration

defaultService :: Service

API version 2021-08-04 of the Amazon Snow Device Management SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by an Amazon Web Services service.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You don't have sufficient access to perform this action.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request would cause a service quota to be exceeded.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

An unexpected error occurred while processing the request.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The request references a resource that doesn't exist.

AttachmentStatus

newtype AttachmentStatus

Constructors

AttachmentStatus'

Fields

fromAttachmentStatus :: Text

Bundled Patterns

pattern AttachmentStatus_ATTACHED :: AttachmentStatus

pattern AttachmentStatus_ATTACHING :: AttachmentStatus

pattern AttachmentStatus_DETACHED :: AttachmentStatus

pattern AttachmentStatus_DETACHING :: AttachmentStatus

ExecutionState

newtype ExecutionState

Constructors

ExecutionState'

Fields

fromExecutionState :: Text

Bundled Patterns

pattern ExecutionState_CANCELED :: ExecutionState

pattern ExecutionState_FAILED :: ExecutionState

pattern ExecutionState_IN_PROGRESS :: ExecutionState

pattern ExecutionState_QUEUED :: ExecutionState

pattern ExecutionState_REJECTED :: ExecutionState

pattern ExecutionState_SUCCEEDED :: ExecutionState

pattern ExecutionState_TIMED_OUT :: ExecutionState

InstanceStateName

newtype InstanceStateName

Constructors

InstanceStateName'

Fields

fromInstanceStateName :: Text

Bundled Patterns

pattern InstanceStateName_PENDING :: InstanceStateName

pattern InstanceStateName_RUNNING :: InstanceStateName

pattern InstanceStateName_SHUTTING_DOWN :: InstanceStateName

pattern InstanceStateName_STOPPED :: InstanceStateName

pattern InstanceStateName_STOPPING :: InstanceStateName

pattern InstanceStateName_TERMINATED :: InstanceStateName

IpAddressAssignment

newtype IpAddressAssignment

Constructors

IpAddressAssignment'

Fields

fromIpAddressAssignment :: Text

Bundled Patterns

pattern IpAddressAssignment_DHCP :: IpAddressAssignment

pattern IpAddressAssignment_STATIC :: IpAddressAssignment

PhysicalConnectorType

newtype PhysicalConnectorType

Constructors

PhysicalConnectorType'

Fields

fromPhysicalConnectorType :: Text

Bundled Patterns

pattern PhysicalConnectorType_QSFP :: PhysicalConnectorType

pattern PhysicalConnectorType_RJ45 :: PhysicalConnectorType

pattern PhysicalConnectorType_RJ45_2 :: PhysicalConnectorType

pattern PhysicalConnectorType_SFP_PLUS :: PhysicalConnectorType

pattern PhysicalConnectorType_WIFI :: PhysicalConnectorType

TaskState

newtype TaskState

Constructors

TaskState'

Fields

fromTaskState :: Text

Bundled Patterns

pattern TaskState_CANCELED :: TaskState

pattern TaskState_COMPLETED :: TaskState

pattern TaskState_IN_PROGRESS :: TaskState

UnlockState

newtype UnlockState

Constructors

UnlockState'

Fields

fromUnlockState :: Text

Bundled Patterns

pattern UnlockState_LOCKED :: UnlockState

pattern UnlockState_UNLOCKED :: UnlockState

pattern UnlockState_UNLOCKING :: UnlockState

Capacity

data Capacity

The physical capacity of the Amazon Web Services Snow Family device.

See: newCapacity smart constructor.

Constructors

Capacity'

Fields

used :: Maybe Integer

The amount of capacity used on the device.

name :: Maybe Text

The name of the type of capacity, such as memory.

total :: Maybe Integer

The total capacity on the device.

unit :: Maybe Text

The unit of measure for the type of capacity.

available :: Maybe Integer

The amount of capacity available for use on the device.

newCapacity :: Capacity

Create a value of Capacity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:used:Capacity', capacity_used - The amount of capacity used on the device.

$sel:name:Capacity', capacity_name - The name of the type of capacity, such as memory.

$sel:total:Capacity', capacity_total - The total capacity on the device.

$sel:unit:Capacity', capacity_unit - The unit of measure for the type of capacity.

$sel:available:Capacity', capacity_available - The amount of capacity available for use on the device.

capacity_used :: Lens' Capacity (Maybe Integer)

The amount of capacity used on the device.

capacity_name :: Lens' Capacity (Maybe Text)

The name of the type of capacity, such as memory.

capacity_total :: Lens' Capacity (Maybe Integer)

The total capacity on the device.

capacity_unit :: Lens' Capacity (Maybe Text)

The unit of measure for the type of capacity.

capacity_available :: Lens' Capacity (Maybe Integer)

The amount of capacity available for use on the device.

Command

data Command

The command given to the device to execute.

See: newCommand smart constructor.

Constructors

Command'

Fields

unlock :: Maybe Unlock

Unlocks the device.

reboot :: Maybe Reboot

Reboots the device.

newCommand :: Command

Create a value of Command with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unlock:Command', command_unlock - Unlocks the device.

$sel:reboot:Command', command_reboot - Reboots the device.

command_unlock :: Lens' Command (Maybe Unlock)

Unlocks the device.

command_reboot :: Lens' Command (Maybe Reboot)

Reboots the device.

CpuOptions

data CpuOptions

The options for how a device's CPU is configured.

See: newCpuOptions smart constructor.

Constructors

CpuOptions'

Fields

coreCount :: Maybe Int

The number of cores that the CPU can use.

threadsPerCore :: Maybe Int

The number of threads per core in the CPU.

newCpuOptions :: CpuOptions

Create a value of CpuOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:coreCount:CpuOptions', cpuOptions_coreCount - The number of cores that the CPU can use.

$sel:threadsPerCore:CpuOptions', cpuOptions_threadsPerCore - The number of threads per core in the CPU.

cpuOptions_coreCount :: Lens' CpuOptions (Maybe Int)

The number of cores that the CPU can use.

cpuOptions_threadsPerCore :: Lens' CpuOptions (Maybe Int)

The number of threads per core in the CPU.

DeviceSummary

data DeviceSummary

Identifying information about the device.

See: newDeviceSummary smart constructor.

Constructors

DeviceSummary'

Fields

associatedWithJob :: Maybe Text

The ID of the job used to order the device.

managedDeviceId :: Maybe Text

The ID of the device.

managedDeviceArn :: Maybe Text

The Amazon Resource Name (ARN) of the device.

tags :: Maybe (HashMap Text Text)

Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

newDeviceSummary :: DeviceSummary

Create a value of DeviceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associatedWithJob:DeviceSummary', deviceSummary_associatedWithJob - The ID of the job used to order the device.

$sel:managedDeviceId:DeviceSummary', deviceSummary_managedDeviceId - The ID of the device.

$sel:managedDeviceArn:DeviceSummary', deviceSummary_managedDeviceArn - The Amazon Resource Name (ARN) of the device.

$sel:tags:DeviceSummary', deviceSummary_tags - Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

deviceSummary_associatedWithJob :: Lens' DeviceSummary (Maybe Text)

The ID of the job used to order the device.

deviceSummary_managedDeviceId :: Lens' DeviceSummary (Maybe Text)

The ID of the device.

deviceSummary_managedDeviceArn :: Lens' DeviceSummary (Maybe Text)

The Amazon Resource Name (ARN) of the device.

deviceSummary_tags :: Lens' DeviceSummary (Maybe (HashMap Text Text))

Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

EbsInstanceBlockDevice

data EbsInstanceBlockDevice

Describes a parameter used to set up an Amazon Elastic Block Store (Amazon EBS) volume in a block device mapping.

See: newEbsInstanceBlockDevice smart constructor.

Constructors

EbsInstanceBlockDevice'

Fields

deleteOnTermination :: Maybe Bool

A value that indicates whether the volume is deleted on instance termination.

status :: Maybe AttachmentStatus

The attachment state.

volumeId :: Maybe Text

The ID of the Amazon EBS volume.

attachTime :: Maybe POSIX

When the attachment was initiated.

newEbsInstanceBlockDevice :: EbsInstanceBlockDevice

Create a value of EbsInstanceBlockDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deleteOnTermination:EbsInstanceBlockDevice', ebsInstanceBlockDevice_deleteOnTermination - A value that indicates whether the volume is deleted on instance termination.

$sel:status:EbsInstanceBlockDevice', ebsInstanceBlockDevice_status - The attachment state.

$sel:volumeId:EbsInstanceBlockDevice', ebsInstanceBlockDevice_volumeId - The ID of the Amazon EBS volume.

$sel:attachTime:EbsInstanceBlockDevice', ebsInstanceBlockDevice_attachTime - When the attachment was initiated.

ebsInstanceBlockDevice_deleteOnTermination :: Lens' EbsInstanceBlockDevice (Maybe Bool)

A value that indicates whether the volume is deleted on instance termination.

ebsInstanceBlockDevice_status :: Lens' EbsInstanceBlockDevice (Maybe AttachmentStatus)

The attachment state.

ebsInstanceBlockDevice_volumeId :: Lens' EbsInstanceBlockDevice (Maybe Text)

The ID of the Amazon EBS volume.

ebsInstanceBlockDevice_attachTime :: Lens' EbsInstanceBlockDevice (Maybe UTCTime)

When the attachment was initiated.

ExecutionSummary

data ExecutionSummary

The summary of a task execution on a specified device.

See: newExecutionSummary smart constructor.

Constructors

ExecutionSummary'

Fields

executionId :: Maybe Text

The ID of the execution.

state :: Maybe ExecutionState

The state of the execution.

taskId :: Maybe Text

The ID of the task.

managedDeviceId :: Maybe Text

The ID of the managed device that the task is being executed on.

newExecutionSummary :: ExecutionSummary

Create a value of ExecutionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:executionId:ExecutionSummary', executionSummary_executionId - The ID of the execution.

$sel:state:ExecutionSummary', executionSummary_state - The state of the execution.

$sel:taskId:ExecutionSummary', executionSummary_taskId - The ID of the task.

$sel:managedDeviceId:ExecutionSummary', executionSummary_managedDeviceId - The ID of the managed device that the task is being executed on.

executionSummary_executionId :: Lens' ExecutionSummary (Maybe Text)

The ID of the execution.

executionSummary_state :: Lens' ExecutionSummary (Maybe ExecutionState)

The state of the execution.

executionSummary_taskId :: Lens' ExecutionSummary (Maybe Text)

The ID of the task.

executionSummary_managedDeviceId :: Lens' ExecutionSummary (Maybe Text)

The ID of the managed device that the task is being executed on.

Instance

data Instance

The description of an instance. Currently, Amazon EC2 instances are the only supported instance type.

See: newInstance smart constructor.

Constructors

Instance'

Fields

instanceId :: Maybe Text

The ID of the instance.

state :: Maybe InstanceState

securityGroups :: Maybe [SecurityGroupIdentifier]

The security groups for the instance.

createdAt :: Maybe POSIX

When the instance was created.

cpuOptions :: Maybe CpuOptions

The CPU options for the instance.

rootDeviceName :: Maybe Text

The device name of the root device volume (for example, /dev/sda1).

instanceType :: Maybe Text

The instance type.

imageId :: Maybe Text

The ID of the AMI used to launch the instance.

privateIpAddress :: Maybe Text

The private IPv4 address assigned to the instance.

updatedAt :: Maybe POSIX

When the instance was last updated.

blockDeviceMappings :: Maybe [InstanceBlockDeviceMapping]

Any block device mapping entries for the instance.

publicIpAddress :: Maybe Text

The public IPv4 address assigned to the instance.

amiLaunchIndex :: Maybe Int

The Amazon Machine Image (AMI) launch index, which you can use to find this instance in the launch group.

newInstance :: Instance

Create a value of Instance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:Instance', instance_instanceId - The ID of the instance.

$sel:state:Instance', instance_state - Undocumented member.

$sel:securityGroups:Instance', instance_securityGroups - The security groups for the instance.

$sel:createdAt:Instance', instance_createdAt - When the instance was created.

$sel:cpuOptions:Instance', instance_cpuOptions - The CPU options for the instance.

$sel:rootDeviceName:Instance', instance_rootDeviceName - The device name of the root device volume (for example, /dev/sda1).

$sel:instanceType:Instance', instance_instanceType - The instance type.

$sel:imageId:Instance', instance_imageId - The ID of the AMI used to launch the instance.

$sel:privateIpAddress:Instance', instance_privateIpAddress - The private IPv4 address assigned to the instance.

$sel:updatedAt:Instance', instance_updatedAt - When the instance was last updated.

$sel:blockDeviceMappings:Instance', instance_blockDeviceMappings - Any block device mapping entries for the instance.

$sel:publicIpAddress:Instance', instance_publicIpAddress - The public IPv4 address assigned to the instance.

$sel:amiLaunchIndex:Instance', instance_amiLaunchIndex - The Amazon Machine Image (AMI) launch index, which you can use to find this instance in the launch group.

instance_instanceId :: Lens' Instance (Maybe Text)

The ID of the instance.

instance_state :: Lens' Instance (Maybe InstanceState)

Undocumented member.

instance_securityGroups :: Lens' Instance (Maybe [SecurityGroupIdentifier])

The security groups for the instance.

instance_createdAt :: Lens' Instance (Maybe UTCTime)

When the instance was created.

instance_cpuOptions :: Lens' Instance (Maybe CpuOptions)

The CPU options for the instance.

instance_rootDeviceName :: Lens' Instance (Maybe Text)

The device name of the root device volume (for example, /dev/sda1).

instance_instanceType :: Lens' Instance (Maybe Text)

The instance type.

instance_imageId :: Lens' Instance (Maybe Text)

The ID of the AMI used to launch the instance.

instance_privateIpAddress :: Lens' Instance (Maybe Text)

The private IPv4 address assigned to the instance.

instance_updatedAt :: Lens' Instance (Maybe UTCTime)

When the instance was last updated.

instance_blockDeviceMappings :: Lens' Instance (Maybe [InstanceBlockDeviceMapping])

Any block device mapping entries for the instance.

instance_publicIpAddress :: Lens' Instance (Maybe Text)

The public IPv4 address assigned to the instance.

instance_amiLaunchIndex :: Lens' Instance (Maybe Int)

The Amazon Machine Image (AMI) launch index, which you can use to find this instance in the launch group.

InstanceBlockDeviceMapping

data InstanceBlockDeviceMapping

The description of a block device mapping.

See: newInstanceBlockDeviceMapping smart constructor.

Constructors

InstanceBlockDeviceMapping'

Fields

ebs :: Maybe EbsInstanceBlockDevice

The parameters used to automatically set up Amazon Elastic Block Store (Amazon EBS) volumes when the instance is launched.

deviceName :: Maybe Text

The block device name.

newInstanceBlockDeviceMapping :: InstanceBlockDeviceMapping

Create a value of InstanceBlockDeviceMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ebs:InstanceBlockDeviceMapping', instanceBlockDeviceMapping_ebs - The parameters used to automatically set up Amazon Elastic Block Store (Amazon EBS) volumes when the instance is launched.

$sel:deviceName:InstanceBlockDeviceMapping', instanceBlockDeviceMapping_deviceName - The block device name.

instanceBlockDeviceMapping_ebs :: Lens' InstanceBlockDeviceMapping (Maybe EbsInstanceBlockDevice)

The parameters used to automatically set up Amazon Elastic Block Store (Amazon EBS) volumes when the instance is launched.

instanceBlockDeviceMapping_deviceName :: Lens' InstanceBlockDeviceMapping (Maybe Text)

The block device name.

InstanceState

data InstanceState

The description of the current state of an instance.

See: newInstanceState smart constructor.

Constructors

InstanceState'

Fields

name :: Maybe InstanceStateName

The current state of the instance.

code :: Maybe Int

The state of the instance as a 16-bit unsigned integer.

The high byte is all of the bits between 2^8 and (2^16)-1, which equals decimal values between 256 and 65,535. These numerical values are used for internal purposes and should be ignored.

The low byte is all of the bits between 2^0 and (2^8)-1, which equals decimal values between 0 and 255.

The valid values for the instance state code are all in the range of the low byte. These values are:

0 : pending

16 : running

32 : shutting-down

48 : terminated

64 : stopping

80 : stopped

You can ignore the high byte value by zeroing out all of the bits above 2^8 or 256 in decimal.

newInstanceState :: InstanceState

Create a value of InstanceState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:InstanceState', instanceState_name - The current state of the instance.

$sel:code:InstanceState', instanceState_code - The state of the instance as a 16-bit unsigned integer.

The high byte is all of the bits between 2^8 and (2^16)-1, which equals decimal values between 256 and 65,535. These numerical values are used for internal purposes and should be ignored.

The low byte is all of the bits between 2^0 and (2^8)-1, which equals decimal values between 0 and 255.

The valid values for the instance state code are all in the range of the low byte. These values are:

0 : pending

16 : running

32 : shutting-down

48 : terminated

64 : stopping

80 : stopped

You can ignore the high byte value by zeroing out all of the bits above 2^8 or 256 in decimal.

instanceState_name :: Lens' InstanceState (Maybe InstanceStateName)

The current state of the instance.

instanceState_code :: Lens' InstanceState (Maybe Int)

The state of the instance as a 16-bit unsigned integer.

The high byte is all of the bits between 2^8 and (2^16)-1, which equals decimal values between 256 and 65,535. These numerical values are used for internal purposes and should be ignored.

The low byte is all of the bits between 2^0 and (2^8)-1, which equals decimal values between 0 and 255.

The valid values for the instance state code are all in the range of the low byte. These values are:

0 : pending

16 : running

32 : shutting-down

48 : terminated

64 : stopping

80 : stopped

You can ignore the high byte value by zeroing out all of the bits above 2^8 or 256 in decimal.

InstanceSummary

data InstanceSummary

The details about the instance.

See: newInstanceSummary smart constructor.

Constructors

InstanceSummary'

Fields

lastUpdatedAt :: Maybe POSIX

When the instance summary was last updated.

instance' :: Maybe Instance

A structure containing details about the instance.

newInstanceSummary :: InstanceSummary

Create a value of InstanceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedAt:InstanceSummary', instanceSummary_lastUpdatedAt - When the instance summary was last updated.

$sel:instance':InstanceSummary', instanceSummary_instance - A structure containing details about the instance.

instanceSummary_lastUpdatedAt :: Lens' InstanceSummary (Maybe UTCTime)

When the instance summary was last updated.

instanceSummary_instance :: Lens' InstanceSummary (Maybe Instance)

A structure containing details about the instance.

PhysicalNetworkInterface

data PhysicalNetworkInterface

The details about the physical network interface for the device.

See: newPhysicalNetworkInterface smart constructor.

Constructors

PhysicalNetworkInterface'

Fields

ipAddress :: Maybe Text

The IP address of the device.

macAddress :: Maybe Text

The MAC address of the device.

ipAddressAssignment :: Maybe IpAddressAssignment

A value that describes whether the IP address is dynamic or persistent.

defaultGateway :: Maybe Text

The default gateway of the device.

physicalNetworkInterfaceId :: Maybe Text

The physical network interface ID.

netmask :: Maybe Text

The netmask used to divide the IP address into subnets.

physicalConnectorType :: Maybe PhysicalConnectorType

The physical connector type.

newPhysicalNetworkInterface :: PhysicalNetworkInterface

Create a value of PhysicalNetworkInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipAddress:PhysicalNetworkInterface', physicalNetworkInterface_ipAddress - The IP address of the device.

$sel:macAddress:PhysicalNetworkInterface', physicalNetworkInterface_macAddress - The MAC address of the device.

$sel:ipAddressAssignment:PhysicalNetworkInterface', physicalNetworkInterface_ipAddressAssignment - A value that describes whether the IP address is dynamic or persistent.

$sel:defaultGateway:PhysicalNetworkInterface', physicalNetworkInterface_defaultGateway - The default gateway of the device.

$sel:physicalNetworkInterfaceId:PhysicalNetworkInterface', physicalNetworkInterface_physicalNetworkInterfaceId - The physical network interface ID.

$sel:netmask:PhysicalNetworkInterface', physicalNetworkInterface_netmask - The netmask used to divide the IP address into subnets.

$sel:physicalConnectorType:PhysicalNetworkInterface', physicalNetworkInterface_physicalConnectorType - The physical connector type.

physicalNetworkInterface_ipAddress :: Lens' PhysicalNetworkInterface (Maybe Text)

The IP address of the device.

physicalNetworkInterface_macAddress :: Lens' PhysicalNetworkInterface (Maybe Text)

The MAC address of the device.

physicalNetworkInterface_ipAddressAssignment :: Lens' PhysicalNetworkInterface (Maybe IpAddressAssignment)

A value that describes whether the IP address is dynamic or persistent.

physicalNetworkInterface_defaultGateway :: Lens' PhysicalNetworkInterface (Maybe Text)

The default gateway of the device.

physicalNetworkInterface_physicalNetworkInterfaceId :: Lens' PhysicalNetworkInterface (Maybe Text)

The physical network interface ID.

physicalNetworkInterface_netmask :: Lens' PhysicalNetworkInterface (Maybe Text)

The netmask used to divide the IP address into subnets.

physicalNetworkInterface_physicalConnectorType :: Lens' PhysicalNetworkInterface (Maybe PhysicalConnectorType)

The physical connector type.

Reboot

data Reboot

A structure used to reboot the device.

See: newReboot smart constructor.

Constructors

Reboot'

newReboot :: Reboot

Create a value of Reboot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

ResourceSummary

data ResourceSummary

A summary of a resource available on the device.

See: newResourceSummary smart constructor.

Constructors

ResourceSummary'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the resource.

id :: Maybe Text

The ID of the resource.

resourceType :: Text

The resource type.

newResourceSummary

Arguments

:: Text
$sel:resourceType:ResourceSummary'

-> ResourceSummary

Create a value of ResourceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ResourceSummary', resourceSummary_arn - The Amazon Resource Name (ARN) of the resource.

$sel:id:ResourceSummary', resourceSummary_id - The ID of the resource.

$sel:resourceType:ResourceSummary', resourceSummary_resourceType - The resource type.

resourceSummary_arn :: Lens' ResourceSummary (Maybe Text)

The Amazon Resource Name (ARN) of the resource.

resourceSummary_id :: Lens' ResourceSummary (Maybe Text)

The ID of the resource.

resourceSummary_resourceType :: Lens' ResourceSummary Text

The resource type.

SecurityGroupIdentifier

data SecurityGroupIdentifier

Information about the device's security group.

See: newSecurityGroupIdentifier smart constructor.

Constructors

SecurityGroupIdentifier'

Fields

groupId :: Maybe Text

The security group ID.

groupName :: Maybe Text

The security group name.

newSecurityGroupIdentifier :: SecurityGroupIdentifier

Create a value of SecurityGroupIdentifier with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupId:SecurityGroupIdentifier', securityGroupIdentifier_groupId - The security group ID.

$sel:groupName:SecurityGroupIdentifier', securityGroupIdentifier_groupName - The security group name.

securityGroupIdentifier_groupId :: Lens' SecurityGroupIdentifier (Maybe Text)

The security group ID.

securityGroupIdentifier_groupName :: Lens' SecurityGroupIdentifier (Maybe Text)

The security group name.

SoftwareInformation

data SoftwareInformation

Information about the software on the device.

See: newSoftwareInformation smart constructor.

Constructors

SoftwareInformation'

Fields

installedVersion :: Maybe Text

The version of the software currently installed on the device.

installState :: Maybe Text

The state of the software that is installed or that is being installed on the device.

installingVersion :: Maybe Text

The version of the software being installed on the device.

newSoftwareInformation :: SoftwareInformation

Create a value of SoftwareInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:installedVersion:SoftwareInformation', softwareInformation_installedVersion - The version of the software currently installed on the device.

$sel:installState:SoftwareInformation', softwareInformation_installState - The state of the software that is installed or that is being installed on the device.

$sel:installingVersion:SoftwareInformation', softwareInformation_installingVersion - The version of the software being installed on the device.

softwareInformation_installedVersion :: Lens' SoftwareInformation (Maybe Text)

The version of the software currently installed on the device.

softwareInformation_installState :: Lens' SoftwareInformation (Maybe Text)

The state of the software that is installed or that is being installed on the device.

softwareInformation_installingVersion :: Lens' SoftwareInformation (Maybe Text)

The version of the software being installed on the device.

TaskSummary

data TaskSummary

Information about the task assigned to one or many devices.

See: newTaskSummary smart constructor.

Constructors

TaskSummary'

Fields

state :: Maybe TaskState

The state of the task assigned to one or many devices.

taskArn :: Maybe Text

The Amazon Resource Name (ARN) of the task.

tags :: Maybe (HashMap Text Text)

Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

taskId :: Text

The task ID.

newTaskSummary

Arguments

:: Text
$sel:taskId:TaskSummary'

-> TaskSummary

Create a value of TaskSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:TaskSummary', taskSummary_state - The state of the task assigned to one or many devices.

$sel:taskArn:TaskSummary', taskSummary_taskArn - The Amazon Resource Name (ARN) of the task.

$sel:tags:TaskSummary', taskSummary_tags - Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

$sel:taskId:TaskSummary', taskSummary_taskId - The task ID.

taskSummary_state :: Lens' TaskSummary (Maybe TaskState)

The state of the task assigned to one or many devices.

taskSummary_taskArn :: Lens' TaskSummary (Maybe Text)

The Amazon Resource Name (ARN) of the task.

taskSummary_tags :: Lens' TaskSummary (Maybe (HashMap Text Text))

Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

taskSummary_taskId :: Lens' TaskSummary Text

The task ID.

Unlock

data Unlock

A structure used to unlock a device.

See: newUnlock smart constructor.

Constructors

Unlock'

newUnlock :: Unlock

Create a value of Unlock with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.