Network.AWS.Snowball.CreateCluster

Description

Creates an empty cluster. Each cluster supports five nodes. You use the CreateJob action separately to create the jobs for each of these nodes. The cluster does not ship until these five node jobs have been created.

Creating a Request

data CreateCluster

See: newCreateCluster smart constructor.

Constructors

CreateCluster'

Fields

kmsKeyARN :: Maybe Text

The KmsKeyARN value that you want to associate with this cluster. KmsKeyARN values are created by using the CreateKey API action in AWS Key Management Service (AWS KMS).

remoteManagement :: Maybe RemoteManagement

Allows you to securely operate and manage Snow devices in a cluster remotely from outside of your internal network. When set to INSTALLED_AUTOSTART, remote management will automatically be available when the device arrives at your location. Otherwise, you need to use the Snowball Client to manage the device.

notification :: Maybe Notification

The Amazon Simple Notification Service (Amazon SNS) notification settings for this cluster.

forwardingAddressId :: Maybe Text

The forwarding address ID for a cluster. This field is not supported in most regions.

onDeviceServiceConfiguration :: Maybe OnDeviceServiceConfiguration

Specifies the service or services on the Snow Family device that your transferred data will be exported from or imported into. AWS Snow Family supports Amazon S3 and NFS (Network File System).

description :: Maybe Text

An optional description of this specific cluster, for example Environmental Data Cluster-01.

taxDocuments :: Maybe TaxDocuments

The tax documents required in your AWS Region.

jobType :: JobType

The type of job for this cluster. Currently, the only job type supported for clusters is LOCAL_USE.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

resources :: JobResource

The resources associated with the cluster job. These resources include Amazon S3 buckets and optional AWS Lambda functions written in the Python language.

addressId :: Text

The ID for the address that you want the cluster shipped to.

roleARN :: Text

The RoleARN that you want to associate with this cluster. RoleArn values are created by using the CreateRole API action in AWS Identity and Access Management (IAM).

snowballType :: SnowballType

The type of AWS Snow Family device to use for this cluster.

For cluster jobs, AWS Snow Family currently supports only the EDGE device type.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

shippingOption :: ShippingOption

The shipping speed for each node in this cluster. This speed doesn't dictate how soon you'll get each Snowball Edge device, rather it represents how quickly each device moves to its destination while in transit. Regional shipping speeds are as follows:

In Australia, you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snow devices are delivered in one to seven days.

In the United States of America (US), you have access to one-day shipping and two-day shipping.

In Australia, you have access to express shipping. Typically, devices shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snow devices are delivered in one to seven days.

In the US, you have access to one-day shipping and two-day shipping.

newCreateCluster

Arguments

:: JobType
$sel:jobType:CreateCluster'

-> JobResource
$sel:resources:CreateCluster'

-> Text
$sel:addressId:CreateCluster'

-> Text
$sel:roleARN:CreateCluster'

-> SnowballType
$sel:snowballType:CreateCluster'

-> ShippingOption
$sel:shippingOption:CreateCluster'

-> CreateCluster

Create a value of CreateCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyARN:CreateCluster', createCluster_kmsKeyARN - The KmsKeyARN value that you want to associate with this cluster. KmsKeyARN values are created by using the CreateKey API action in AWS Key Management Service (AWS KMS).

$sel:remoteManagement:CreateCluster', createCluster_remoteManagement - Allows you to securely operate and manage Snow devices in a cluster remotely from outside of your internal network. When set to INSTALLED_AUTOSTART, remote management will automatically be available when the device arrives at your location. Otherwise, you need to use the Snowball Client to manage the device.

$sel:notification:CreateCluster', createCluster_notification - The Amazon Simple Notification Service (Amazon SNS) notification settings for this cluster.

$sel:forwardingAddressId:CreateCluster', createCluster_forwardingAddressId - The forwarding address ID for a cluster. This field is not supported in most regions.

$sel:onDeviceServiceConfiguration:CreateCluster', createCluster_onDeviceServiceConfiguration - Specifies the service or services on the Snow Family device that your transferred data will be exported from or imported into. AWS Snow Family supports Amazon S3 and NFS (Network File System).

$sel:description:CreateCluster', createCluster_description - An optional description of this specific cluster, for example Environmental Data Cluster-01.

$sel:taxDocuments:CreateCluster', createCluster_taxDocuments - The tax documents required in your AWS Region.

$sel:jobType:CreateCluster', createCluster_jobType - The type of job for this cluster. Currently, the only job type supported for clusters is LOCAL_USE.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

$sel:resources:CreateCluster', createCluster_resources - The resources associated with the cluster job. These resources include Amazon S3 buckets and optional AWS Lambda functions written in the Python language.

$sel:addressId:CreateCluster', createCluster_addressId - The ID for the address that you want the cluster shipped to.

$sel:roleARN:CreateCluster', createCluster_roleARN - The RoleARN that you want to associate with this cluster. RoleArn values are created by using the CreateRole API action in AWS Identity and Access Management (IAM).

$sel:snowballType:CreateCluster', createCluster_snowballType - The type of AWS Snow Family device to use for this cluster.

For cluster jobs, AWS Snow Family currently supports only the EDGE device type.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

$sel:shippingOption:CreateCluster', createCluster_shippingOption - The shipping speed for each node in this cluster. This speed doesn't dictate how soon you'll get each Snowball Edge device, rather it represents how quickly each device moves to its destination while in transit. Regional shipping speeds are as follows:

In Australia, you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snow devices are delivered in one to seven days.

In the United States of America (US), you have access to one-day shipping and two-day shipping.

In Australia, you have access to express shipping. Typically, devices shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snow devices are delivered in one to seven days.

In the US, you have access to one-day shipping and two-day shipping.

Request Lenses

createCluster_kmsKeyARN :: Lens' CreateCluster (Maybe Text)

The KmsKeyARN value that you want to associate with this cluster. KmsKeyARN values are created by using the CreateKey API action in AWS Key Management Service (AWS KMS).

createCluster_remoteManagement :: Lens' CreateCluster (Maybe RemoteManagement)

Allows you to securely operate and manage Snow devices in a cluster remotely from outside of your internal network. When set to INSTALLED_AUTOSTART, remote management will automatically be available when the device arrives at your location. Otherwise, you need to use the Snowball Client to manage the device.

createCluster_notification :: Lens' CreateCluster (Maybe Notification)

The Amazon Simple Notification Service (Amazon SNS) notification settings for this cluster.

createCluster_forwardingAddressId :: Lens' CreateCluster (Maybe Text)

The forwarding address ID for a cluster. This field is not supported in most regions.

createCluster_onDeviceServiceConfiguration :: Lens' CreateCluster (Maybe OnDeviceServiceConfiguration)

Specifies the service or services on the Snow Family device that your transferred data will be exported from or imported into. AWS Snow Family supports Amazon S3 and NFS (Network File System).

createCluster_description :: Lens' CreateCluster (Maybe Text)

An optional description of this specific cluster, for example Environmental Data Cluster-01.

createCluster_taxDocuments :: Lens' CreateCluster (Maybe TaxDocuments)

The tax documents required in your AWS Region.

createCluster_jobType :: Lens' CreateCluster JobType

The type of job for this cluster. Currently, the only job type supported for clusters is LOCAL_USE.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

createCluster_resources :: Lens' CreateCluster JobResource

The resources associated with the cluster job. These resources include Amazon S3 buckets and optional AWS Lambda functions written in the Python language.

createCluster_addressId :: Lens' CreateCluster Text

The ID for the address that you want the cluster shipped to.

createCluster_roleARN :: Lens' CreateCluster Text

The RoleARN that you want to associate with this cluster. RoleArn values are created by using the CreateRole API action in AWS Identity and Access Management (IAM).

createCluster_snowballType :: Lens' CreateCluster SnowballType

The type of AWS Snow Family device to use for this cluster.

For cluster jobs, AWS Snow Family currently supports only the EDGE device type.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

createCluster_shippingOption :: Lens' CreateCluster ShippingOption

The shipping speed for each node in this cluster. This speed doesn't dictate how soon you'll get each Snowball Edge device, rather it represents how quickly each device moves to its destination while in transit. Regional shipping speeds are as follows:

In Australia, you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snow devices are delivered in one to seven days.

In the United States of America (US), you have access to one-day shipping and two-day shipping.

In Australia, you have access to express shipping. Typically, devices shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snow devices are delivered in one to seven days.

In the US, you have access to one-day shipping and two-day shipping.

Destructuring the Response

data CreateClusterResponse

See: newCreateClusterResponse smart constructor.

Constructors

CreateClusterResponse'

Fields

clusterId :: Maybe Text

The automatically generated ID for a cluster.

httpStatus :: Int

The response's http status code.

newCreateClusterResponse

Arguments

:: Int
$sel:httpStatus:CreateClusterResponse'

-> CreateClusterResponse

Create a value of CreateClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterId:CreateClusterResponse', createClusterResponse_clusterId - The automatically generated ID for a cluster.

$sel:httpStatus:CreateClusterResponse', createClusterResponse_httpStatus - The response's http status code.

Response Lenses

createClusterResponse_clusterId :: Lens' CreateClusterResponse (Maybe Text)

The automatically generated ID for a cluster.

createClusterResponse_httpStatus :: Lens' CreateClusterResponse Int

The response's http status code.