Network.AWS.Snowball.CreateJob

Description

Creates a job to import or export data between Amazon S3 and your on-premises data center. Your AWS account must have the right trust policies and permissions in place to create a job for a Snow device. If you're creating a job for a node in a cluster, you only need to provide the clusterId value; the other job attributes are inherited from the cluster.

Only the Snowball; Edge device type is supported when ordering clustered jobs.

The device capacity is optional.

Availability of device types differ by AWS Region. For more information about Region availability, see AWS Regional Services.

AWS Snow Family device types and their capacities.

Snow Family device type: SNC1_SSD

Capacity: T14

Description: Snowcone

Snow Family device type: SNC1_HDD

Capacity: T8

Description: Snowcone

Device type: EDGE_S

Capacity: T98

Description: Snowball Edge Storage Optimized for data transfer only

Device type: EDGE_CG

Capacity: T42

Description: Snowball Edge Compute Optimized with GPU

Device type: EDGE_C

Capacity: T42

Description: Snowball Edge Compute Optimized without GPU

Device type: EDGE

Capacity: T100

Description: Snowball Edge Storage Optimized with EC2 Compute

Device type: STANDARD

Capacity: T50

Description: Original Snowball device

This device is only available in the Ningxia, Beijing, and Singapore AWS Regions.

Device type: STANDARD

Capacity: T80

Description: Original Snowball device

This device is only available in the Ningxia, Beijing, and Singapore AWS Regions.

Creating a Request

data CreateJob

See: newCreateJob smart constructor.

Constructors

CreateJob'

Fields

jobType :: Maybe JobType

Defines the type of job that you're creating.

kmsKeyARN :: Maybe Text

The KmsKeyARN that you want to associate with this job. KmsKeyARNs are created using the CreateKey AWS Key Management Service (KMS) API action.

remoteManagement :: Maybe RemoteManagement

Allows you to securely operate and manage Snowcone devices remotely from outside of your internal network. When set to INSTALLED_AUTOSTART, remote management will automatically be available when the device arrives at your location. Otherwise, you need to use the Snowball Client to manage the device.

notification :: Maybe Notification

Defines the Amazon Simple Notification Service (Amazon SNS) notification settings for this job.

forwardingAddressId :: Maybe Text

The forwarding address ID for a job. This field is not supported in most Regions.

addressId :: Maybe Text

The ID for the address that you want the Snow device shipped to.

snowballType :: Maybe SnowballType

The type of AWS Snow Family device to use for this job.

For cluster jobs, AWS Snow Family currently supports only the EDGE device type.

The type of AWS Snow device to use for this job. Currently, the only supported device type for cluster jobs is EDGE.

For more information, see Snowball Edge Device Options in the Snowball Edge Developer Guide.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

longTermPricingId :: Maybe Text

The ID of the long-term pricing type for the device.

shippingOption :: Maybe ShippingOption

The shipping speed for this job. This speed doesn't dictate how soon you'll get the Snow device, rather it represents how quickly the Snow device moves to its destination while in transit. Regional shipping speeds are as follows:

In Australia, you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snow devices are delivered in one to seven days.

In the US, you have access to one-day shipping and two-day shipping.

resources :: Maybe JobResource

Defines the Amazon S3 buckets associated with this job.

With IMPORT jobs, you specify the bucket or buckets that your transferred data will be imported into.

With EXPORT jobs, you specify the bucket or buckets that your transferred data will be exported from. Optionally, you can also specify a KeyRange value. If you choose to export a range, you define the length of the range by providing either an inclusive BeginMarker value, an inclusive EndMarker value, or both. Ranges are UTF-8 binary sorted.

onDeviceServiceConfiguration :: Maybe OnDeviceServiceConfiguration

Specifies the service or services on the Snow Family device that your transferred data will be exported from or imported into. AWS Snow Family supports Amazon S3 and NFS (Network File System).

clusterId :: Maybe Text

The ID of a cluster. If you're creating a job for a node in a cluster, you need to provide only this clusterId value. The other job attributes are inherited from the cluster.

deviceConfiguration :: Maybe DeviceConfiguration

Defines the device configuration for an AWS Snowcone job.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

description :: Maybe Text

Defines an optional description of this specific job, for example Important Photos 2016-08-11.

taxDocuments :: Maybe TaxDocuments

The tax documents required in your AWS Region.

roleARN :: Maybe Text

The RoleARN that you want to associate with this job. RoleArns are created using the CreateRole AWS Identity and Access Management (IAM) API action.

snowballCapacityPreference :: Maybe SnowballCapacity

If your job is being created in one of the US regions, you have the option of specifying what size Snow device you'd like for this job. In all other regions, Snowballs come with 80 TB in storage capacity.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

newCreateJob :: CreateJob

Create a value of CreateJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobType:CreateJob', createJob_jobType - Defines the type of job that you're creating.

$sel:kmsKeyARN:CreateJob', createJob_kmsKeyARN - The KmsKeyARN that you want to associate with this job. KmsKeyARNs are created using the CreateKey AWS Key Management Service (KMS) API action.

$sel:remoteManagement:CreateJob', createJob_remoteManagement - Allows you to securely operate and manage Snowcone devices remotely from outside of your internal network. When set to INSTALLED_AUTOSTART, remote management will automatically be available when the device arrives at your location. Otherwise, you need to use the Snowball Client to manage the device.

$sel:notification:CreateJob', createJob_notification - Defines the Amazon Simple Notification Service (Amazon SNS) notification settings for this job.

$sel:forwardingAddressId:CreateJob', createJob_forwardingAddressId - The forwarding address ID for a job. This field is not supported in most Regions.

$sel:addressId:CreateJob', createJob_addressId - The ID for the address that you want the Snow device shipped to.

$sel:snowballType:CreateJob', createJob_snowballType - The type of AWS Snow Family device to use for this job.

For cluster jobs, AWS Snow Family currently supports only the EDGE device type.

The type of AWS Snow device to use for this job. Currently, the only supported device type for cluster jobs is EDGE.

For more information, see Snowball Edge Device Options in the Snowball Edge Developer Guide.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

$sel:longTermPricingId:CreateJob', createJob_longTermPricingId - The ID of the long-term pricing type for the device.

$sel:shippingOption:CreateJob', createJob_shippingOption - The shipping speed for this job. This speed doesn't dictate how soon you'll get the Snow device, rather it represents how quickly the Snow device moves to its destination while in transit. Regional shipping speeds are as follows:

In Australia, you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snow devices are delivered in one to seven days.

In the US, you have access to one-day shipping and two-day shipping.

$sel:resources:CreateJob', createJob_resources - Defines the Amazon S3 buckets associated with this job.

With IMPORT jobs, you specify the bucket or buckets that your transferred data will be imported into.

With EXPORT jobs, you specify the bucket or buckets that your transferred data will be exported from. Optionally, you can also specify a KeyRange value. If you choose to export a range, you define the length of the range by providing either an inclusive BeginMarker value, an inclusive EndMarker value, or both. Ranges are UTF-8 binary sorted.

$sel:onDeviceServiceConfiguration:CreateJob', createJob_onDeviceServiceConfiguration - Specifies the service or services on the Snow Family device that your transferred data will be exported from or imported into. AWS Snow Family supports Amazon S3 and NFS (Network File System).

$sel:clusterId:CreateJob', createJob_clusterId - The ID of a cluster. If you're creating a job for a node in a cluster, you need to provide only this clusterId value. The other job attributes are inherited from the cluster.

$sel:deviceConfiguration:CreateJob', createJob_deviceConfiguration - Defines the device configuration for an AWS Snowcone job.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

$sel:description:CreateJob', createJob_description - Defines an optional description of this specific job, for example Important Photos 2016-08-11.

$sel:taxDocuments:CreateJob', createJob_taxDocuments - The tax documents required in your AWS Region.

$sel:roleARN:CreateJob', createJob_roleARN - The RoleARN that you want to associate with this job. RoleArns are created using the CreateRole AWS Identity and Access Management (IAM) API action.

$sel:snowballCapacityPreference:CreateJob', createJob_snowballCapacityPreference - If your job is being created in one of the US regions, you have the option of specifying what size Snow device you'd like for this job. In all other regions, Snowballs come with 80 TB in storage capacity.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

Request Lenses

createJob_jobType :: Lens' CreateJob (Maybe JobType)

Defines the type of job that you're creating.

createJob_kmsKeyARN :: Lens' CreateJob (Maybe Text)

The KmsKeyARN that you want to associate with this job. KmsKeyARNs are created using the CreateKey AWS Key Management Service (KMS) API action.

createJob_remoteManagement :: Lens' CreateJob (Maybe RemoteManagement)

Allows you to securely operate and manage Snowcone devices remotely from outside of your internal network. When set to INSTALLED_AUTOSTART, remote management will automatically be available when the device arrives at your location. Otherwise, you need to use the Snowball Client to manage the device.

createJob_notification :: Lens' CreateJob (Maybe Notification)

Defines the Amazon Simple Notification Service (Amazon SNS) notification settings for this job.

createJob_forwardingAddressId :: Lens' CreateJob (Maybe Text)

The forwarding address ID for a job. This field is not supported in most Regions.

createJob_addressId :: Lens' CreateJob (Maybe Text)

The ID for the address that you want the Snow device shipped to.

createJob_snowballType :: Lens' CreateJob (Maybe SnowballType)

The type of AWS Snow Family device to use for this job.

For cluster jobs, AWS Snow Family currently supports only the EDGE device type.

The type of AWS Snow device to use for this job. Currently, the only supported device type for cluster jobs is EDGE.

For more information, see Snowball Edge Device Options in the Snowball Edge Developer Guide.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

createJob_longTermPricingId :: Lens' CreateJob (Maybe Text)

The ID of the long-term pricing type for the device.

createJob_shippingOption :: Lens' CreateJob (Maybe ShippingOption)

The shipping speed for this job. This speed doesn't dictate how soon you'll get the Snow device, rather it represents how quickly the Snow device moves to its destination while in transit. Regional shipping speeds are as follows:

In Australia, you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snow devices are delivered in one to seven days.

In the US, you have access to one-day shipping and two-day shipping.

createJob_resources :: Lens' CreateJob (Maybe JobResource)

Defines the Amazon S3 buckets associated with this job.

With IMPORT jobs, you specify the bucket or buckets that your transferred data will be imported into.

With EXPORT jobs, you specify the bucket or buckets that your transferred data will be exported from. Optionally, you can also specify a KeyRange value. If you choose to export a range, you define the length of the range by providing either an inclusive BeginMarker value, an inclusive EndMarker value, or both. Ranges are UTF-8 binary sorted.

createJob_onDeviceServiceConfiguration :: Lens' CreateJob (Maybe OnDeviceServiceConfiguration)

Specifies the service or services on the Snow Family device that your transferred data will be exported from or imported into. AWS Snow Family supports Amazon S3 and NFS (Network File System).

createJob_clusterId :: Lens' CreateJob (Maybe Text)

The ID of a cluster. If you're creating a job for a node in a cluster, you need to provide only this clusterId value. The other job attributes are inherited from the cluster.

createJob_deviceConfiguration :: Lens' CreateJob (Maybe DeviceConfiguration)

Defines the device configuration for an AWS Snowcone job.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

createJob_description :: Lens' CreateJob (Maybe Text)

Defines an optional description of this specific job, for example Important Photos 2016-08-11.

createJob_taxDocuments :: Lens' CreateJob (Maybe TaxDocuments)

The tax documents required in your AWS Region.

createJob_roleARN :: Lens' CreateJob (Maybe Text)

The RoleARN that you want to associate with this job. RoleArns are created using the CreateRole AWS Identity and Access Management (IAM) API action.

createJob_snowballCapacityPreference :: Lens' CreateJob (Maybe SnowballCapacity)

If your job is being created in one of the US regions, you have the option of specifying what size Snow device you'd like for this job. In all other regions, Snowballs come with 80 TB in storage capacity.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

Destructuring the Response

data CreateJobResponse

See: newCreateJobResponse smart constructor.

Constructors

CreateJobResponse'

Fields

jobId :: Maybe Text

The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000.

httpStatus :: Int

The response's http status code.

newCreateJobResponse

Arguments

:: Int
$sel:httpStatus:CreateJobResponse'

-> CreateJobResponse

Create a value of CreateJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:CreateJobResponse', createJobResponse_jobId - The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000.

$sel:httpStatus:CreateJobResponse', createJobResponse_httpStatus - The response's http status code.

Response Lenses

createJobResponse_jobId :: Lens' CreateJobResponse (Maybe Text)

The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000.

createJobResponse_httpStatus :: Lens' CreateJobResponse Int

The response's http status code.