Network.AWS.Snowball.CreateLongTermPricing

Description

Creates a job with the long-term usage option for a device. The long-term usage is a 1-year or 3-year long-term pricing type for the device. You are billed upfront, and AWS provides discounts for long-term pricing.

Creating a Request

data CreateLongTermPricing

See: newCreateLongTermPricing smart constructor.

Constructors

CreateLongTermPricing'

Fields

snowballType :: Maybe SnowballType

The type of AWS Snow Family device to use for the long-term pricing job.

isLongTermPricingAutoRenew :: Maybe Bool

Specifies whether the current long-term pricing type for the device should be renewed.

longTermPricingType :: LongTermPricingType

The type of long-term pricing option you want for the device, either 1-year or 3-year long-term pricing.

newCreateLongTermPricing

Arguments

:: LongTermPricingType
$sel:longTermPricingType:CreateLongTermPricing'

-> CreateLongTermPricing

Create a value of CreateLongTermPricing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snowballType:CreateLongTermPricing', createLongTermPricing_snowballType - The type of AWS Snow Family device to use for the long-term pricing job.

$sel:isLongTermPricingAutoRenew:CreateLongTermPricing', createLongTermPricing_isLongTermPricingAutoRenew - Specifies whether the current long-term pricing type for the device should be renewed.

$sel:longTermPricingType:CreateLongTermPricing', createLongTermPricing_longTermPricingType - The type of long-term pricing option you want for the device, either 1-year or 3-year long-term pricing.

Request Lenses

createLongTermPricing_snowballType :: Lens' CreateLongTermPricing (Maybe SnowballType)

The type of AWS Snow Family device to use for the long-term pricing job.

createLongTermPricing_isLongTermPricingAutoRenew :: Lens' CreateLongTermPricing (Maybe Bool)

Specifies whether the current long-term pricing type for the device should be renewed.

createLongTermPricing_longTermPricingType :: Lens' CreateLongTermPricing LongTermPricingType

The type of long-term pricing option you want for the device, either 1-year or 3-year long-term pricing.

Destructuring the Response

data CreateLongTermPricingResponse

See: newCreateLongTermPricingResponse smart constructor.

Constructors

CreateLongTermPricingResponse'

Fields

longTermPricingId :: Maybe Text

The ID of the long-term pricing type for the device.

httpStatus :: Int

The response's http status code.

newCreateLongTermPricingResponse

Arguments

:: Int
$sel:httpStatus:CreateLongTermPricingResponse'

-> CreateLongTermPricingResponse

Create a value of CreateLongTermPricingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:longTermPricingId:CreateLongTermPricingResponse', createLongTermPricingResponse_longTermPricingId - The ID of the long-term pricing type for the device.

$sel:httpStatus:CreateLongTermPricingResponse', createLongTermPricingResponse_httpStatus - The response's http status code.

Response Lenses

createLongTermPricingResponse_longTermPricingId :: Lens' CreateLongTermPricingResponse (Maybe Text)

The ID of the long-term pricing type for the device.

createLongTermPricingResponse_httpStatus :: Lens' CreateLongTermPricingResponse Int

The response's http status code.