Network.AWS.Snowball.CreateReturnShippingLabel

Description

Creates a shipping label that will be used to return the Snow device to AWS.

Creating a Request

data CreateReturnShippingLabel

See: newCreateReturnShippingLabel smart constructor.

Constructors

CreateReturnShippingLabel'

Fields

shippingOption :: Maybe ShippingOption

The shipping speed for a particular job. This speed doesn't dictate how soon the device is returned to AWS. This speed represents how quickly it moves to its destination while in transit. Regional shipping speeds are as follows:

jobId :: Text

The ID for a job that you want to create the return shipping label for; for example, JID123e4567-e89b-12d3-a456-426655440000.

newCreateReturnShippingLabel

Arguments

:: Text
$sel:jobId:CreateReturnShippingLabel'

-> CreateReturnShippingLabel

Create a value of CreateReturnShippingLabel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shippingOption:CreateReturnShippingLabel', createReturnShippingLabel_shippingOption - The shipping speed for a particular job. This speed doesn't dictate how soon the device is returned to AWS. This speed represents how quickly it moves to its destination while in transit. Regional shipping speeds are as follows:

$sel:jobId:CreateReturnShippingLabel', createReturnShippingLabel_jobId - The ID for a job that you want to create the return shipping label for; for example, JID123e4567-e89b-12d3-a456-426655440000.

Request Lenses

createReturnShippingLabel_shippingOption :: Lens' CreateReturnShippingLabel (Maybe ShippingOption)

The shipping speed for a particular job. This speed doesn't dictate how soon the device is returned to AWS. This speed represents how quickly it moves to its destination while in transit. Regional shipping speeds are as follows:

createReturnShippingLabel_jobId :: Lens' CreateReturnShippingLabel Text

The ID for a job that you want to create the return shipping label for; for example, JID123e4567-e89b-12d3-a456-426655440000.

Destructuring the Response

data CreateReturnShippingLabelResponse

See: newCreateReturnShippingLabelResponse smart constructor.

Constructors

CreateReturnShippingLabelResponse'

Fields

status :: Maybe ShippingLabelStatus

The status information of the task on a Snow device that is being returned to AWS.

httpStatus :: Int

The response's http status code.

newCreateReturnShippingLabelResponse

Arguments

:: Int
$sel:httpStatus:CreateReturnShippingLabelResponse'

-> CreateReturnShippingLabelResponse

Create a value of CreateReturnShippingLabelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateReturnShippingLabelResponse', createReturnShippingLabelResponse_status - The status information of the task on a Snow device that is being returned to AWS.

$sel:httpStatus:CreateReturnShippingLabelResponse', createReturnShippingLabelResponse_httpStatus - The response's http status code.

Response Lenses

createReturnShippingLabelResponse_status :: Lens' CreateReturnShippingLabelResponse (Maybe ShippingLabelStatus)

The status information of the task on a Snow device that is being returned to AWS.

createReturnShippingLabelResponse_httpStatus :: Lens' CreateReturnShippingLabelResponse Int

The response's http status code.