Network.AWS.Snowball.GetJobManifest

Description

Returns a link to an Amazon S3 presigned URL for the manifest file associated with the specified JobId value. You can access the manifest file for up to 60 minutes after this request has been made. To access the manifest file after 60 minutes have passed, you'll have to make another call to the GetJobManifest action.

The manifest is an encrypted file that you can download after your job enters the WithCustomer status. The manifest is decrypted by using the UnlockCode code value, when you pass both values to the Snow device through the Snowball client when the client is started for the first time.

As a best practice, we recommend that you don't save a copy of an UnlockCode value in the same location as the manifest file for that job. Saving these separately helps prevent unauthorized parties from gaining access to the Snow device associated with that job.

The credentials of a given job, including its manifest file and unlock code, expire 360 days after the job is created.

Creating a Request

data GetJobManifest

See: newGetJobManifest smart constructor.

Constructors

GetJobManifest'

Fields

jobId :: Text

The ID for a job that you want to get the manifest file for, for example JID123e4567-e89b-12d3-a456-426655440000.

newGetJobManifest

Arguments

:: Text
$sel:jobId:GetJobManifest'

-> GetJobManifest

Create a value of GetJobManifest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:GetJobManifest', getJobManifest_jobId - The ID for a job that you want to get the manifest file for, for example JID123e4567-e89b-12d3-a456-426655440000.

Request Lenses

getJobManifest_jobId :: Lens' GetJobManifest Text

The ID for a job that you want to get the manifest file for, for example JID123e4567-e89b-12d3-a456-426655440000.

Destructuring the Response

data GetJobManifestResponse

See: newGetJobManifestResponse smart constructor.

Constructors

GetJobManifestResponse'

Fields

manifestURI :: Maybe Text

The Amazon S3 presigned URL for the manifest file associated with the specified JobId value.

httpStatus :: Int

The response's http status code.

newGetJobManifestResponse

Arguments

:: Int
$sel:httpStatus:GetJobManifestResponse'

-> GetJobManifestResponse

Create a value of GetJobManifestResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:manifestURI:GetJobManifestResponse', getJobManifestResponse_manifestURI - The Amazon S3 presigned URL for the manifest file associated with the specified JobId value.

$sel:httpStatus:GetJobManifestResponse', getJobManifestResponse_httpStatus - The response's http status code.

Response Lenses

getJobManifestResponse_manifestURI :: Lens' GetJobManifestResponse (Maybe Text)

The Amazon S3 presigned URL for the manifest file associated with the specified JobId value.

getJobManifestResponse_httpStatus :: Lens' GetJobManifestResponse Int

The response's http status code.