Network.AWS.Snowball.ListJobs

Description

Returns an array of JobListEntry objects of the specified length. Each JobListEntry object contains a job's state, a job's ID, and a value that indicates whether the job is a job part, in the case of export jobs. Calling this API action in one of the US regions will return jobs from the list of all jobs associated with this account in all US regions.

This operation returns paginated results.

Creating a Request

data ListJobs

See: newListJobs smart constructor.

Constructors

ListJobs'

Fields

nextToken :: Maybe Text

HTTP requests are stateless. To identify what object comes "next" in the list of JobListEntry objects, you have the option of specifying NextToken as the starting point for your returned list.

maxResults :: Maybe Natural

The number of JobListEntry objects to return.

newListJobs :: ListJobs

Create a value of ListJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListJobs', listJobs_nextToken - HTTP requests are stateless. To identify what object comes "next" in the list of JobListEntry objects, you have the option of specifying NextToken as the starting point for your returned list.

$sel:maxResults:ListJobs', listJobs_maxResults - The number of JobListEntry objects to return.

Request Lenses

listJobs_nextToken :: Lens' ListJobs (Maybe Text)

HTTP requests are stateless. To identify what object comes "next" in the list of JobListEntry objects, you have the option of specifying NextToken as the starting point for your returned list.

listJobs_maxResults :: Lens' ListJobs (Maybe Natural)

The number of JobListEntry objects to return.

Destructuring the Response

data ListJobsResponse

See: newListJobsResponse smart constructor.

Constructors

ListJobsResponse'

Fields

jobListEntries :: Maybe [JobListEntry]

Each JobListEntry object contains a job's state, a job's ID, and a value that indicates whether the job is a job part, in the case of export jobs.

nextToken :: Maybe Text

HTTP requests are stateless. If you use this automatically generated NextToken value in your next ListJobs call, your returned JobListEntry objects will start from this point in the array.

httpStatus :: Int

The response's http status code.

newListJobsResponse

Arguments

:: Int
$sel:httpStatus:ListJobsResponse'

-> ListJobsResponse

Create a value of ListJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobListEntries:ListJobsResponse', listJobsResponse_jobListEntries - Each JobListEntry object contains a job's state, a job's ID, and a value that indicates whether the job is a job part, in the case of export jobs.

$sel:nextToken:ListJobs', listJobsResponse_nextToken - HTTP requests are stateless. If you use this automatically generated NextToken value in your next ListJobs call, your returned JobListEntry objects will start from this point in the array.

$sel:httpStatus:ListJobsResponse', listJobsResponse_httpStatus - The response's http status code.

Response Lenses

listJobsResponse_jobListEntries :: Lens' ListJobsResponse (Maybe [JobListEntry])

Each JobListEntry object contains a job's state, a job's ID, and a value that indicates whether the job is a job part, in the case of export jobs.

listJobsResponse_nextToken :: Lens' ListJobsResponse (Maybe Text)

HTTP requests are stateless. If you use this automatically generated NextToken value in your next ListJobs call, your returned JobListEntry objects will start from this point in the array.

listJobsResponse_httpStatus :: Lens' ListJobsResponse Int

The response's http status code.