Network.AWS.Snowball.Types.Address

Description

Documentation

data Address

The address that you want the Snow device(s) associated with a specific job to be shipped to. Addresses are validated at the time of creation. The address you provide must be located within the serviceable area of your region. Although no individual elements of the Address are required, if the address is invalid or unsupported, then an exception is thrown.

See: newAddress smart constructor.

Constructors

Address'

Fields

company :: Maybe Text

The name of the company to receive a Snow device at an address.

isRestricted :: Maybe Bool

If the address you are creating is a primary address, then set this option to true. This field is not supported in most regions.

phoneNumber :: Maybe Text

The phone number associated with an address that a Snow device is to be delivered to.

postalCode :: Maybe Text

The postal code in an address that a Snow device is to be delivered to.

street1 :: Maybe Text

The first line in a street address that a Snow device is to be delivered to.

landmark :: Maybe Text

This field is no longer used and the value is ignored.

name :: Maybe Text

The name of a person to receive a Snow device at an address.

city :: Maybe Text

The city in an address that a Snow device is to be delivered to.

street2 :: Maybe Text

The second line in a street address that a Snow device is to be delivered to.

addressId :: Maybe Text

The unique ID for an address.

country :: Maybe Text

The country in an address that a Snow device is to be delivered to.

stateOrProvince :: Maybe Text

The state or province in an address that a Snow device is to be delivered to.

prefectureOrDistrict :: Maybe Text

This field is no longer used and the value is ignored.

street3 :: Maybe Text

The third line in a street address that a Snow device is to be delivered to.

newAddress :: Address

Create a value of Address with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:company:Address', address_company - The name of the company to receive a Snow device at an address.

$sel:isRestricted:Address', address_isRestricted - If the address you are creating is a primary address, then set this option to true. This field is not supported in most regions.

$sel:phoneNumber:Address', address_phoneNumber - The phone number associated with an address that a Snow device is to be delivered to.

$sel:postalCode:Address', address_postalCode - The postal code in an address that a Snow device is to be delivered to.

$sel:street1:Address', address_street1 - The first line in a street address that a Snow device is to be delivered to.

$sel:landmark:Address', address_landmark - This field is no longer used and the value is ignored.

$sel:name:Address', address_name - The name of a person to receive a Snow device at an address.

$sel:city:Address', address_city - The city in an address that a Snow device is to be delivered to.

$sel:street2:Address', address_street2 - The second line in a street address that a Snow device is to be delivered to.

$sel:addressId:Address', address_addressId - The unique ID for an address.

$sel:country:Address', address_country - The country in an address that a Snow device is to be delivered to.

$sel:stateOrProvince:Address', address_stateOrProvince - The state or province in an address that a Snow device is to be delivered to.

$sel:prefectureOrDistrict:Address', address_prefectureOrDistrict - This field is no longer used and the value is ignored.

$sel:street3:Address', address_street3 - The third line in a street address that a Snow device is to be delivered to.

address_company :: Lens' Address (Maybe Text)

The name of the company to receive a Snow device at an address.

address_isRestricted :: Lens' Address (Maybe Bool)

If the address you are creating is a primary address, then set this option to true. This field is not supported in most regions.

address_phoneNumber :: Lens' Address (Maybe Text)

The phone number associated with an address that a Snow device is to be delivered to.

address_postalCode :: Lens' Address (Maybe Text)

The postal code in an address that a Snow device is to be delivered to.

address_street1 :: Lens' Address (Maybe Text)

The first line in a street address that a Snow device is to be delivered to.

address_landmark :: Lens' Address (Maybe Text)

This field is no longer used and the value is ignored.

address_name :: Lens' Address (Maybe Text)

The name of a person to receive a Snow device at an address.

address_city :: Lens' Address (Maybe Text)

The city in an address that a Snow device is to be delivered to.

address_street2 :: Lens' Address (Maybe Text)

The second line in a street address that a Snow device is to be delivered to.

address_addressId :: Lens' Address (Maybe Text)

The unique ID for an address.

address_country :: Lens' Address (Maybe Text)

The country in an address that a Snow device is to be delivered to.

address_stateOrProvince :: Lens' Address (Maybe Text)

The state or province in an address that a Snow device is to be delivered to.

address_prefectureOrDistrict :: Lens' Address (Maybe Text)

This field is no longer used and the value is ignored.

address_street3 :: Lens' Address (Maybe Text)

The third line in a street address that a Snow device is to be delivered to.