Network.AWS.Snowball.Types.ClusterListEntry

Description

Documentation

data ClusterListEntry

Contains a cluster's state, a cluster's ID, and other important information.

See: newClusterListEntry smart constructor.

Constructors

ClusterListEntry'

Fields

clusterState :: Maybe ClusterState

The current state of this cluster. For information about the state of a specific node, see JobListEntry$JobState.

clusterId :: Maybe Text

The 39-character ID for the cluster that you want to list, for example CID123e4567-e89b-12d3-a456-426655440000.

creationDate :: Maybe POSIX

The creation date for this cluster.

description :: Maybe Text

Defines an optional description of the cluster, for example Environmental Data Cluster-01.

newClusterListEntry :: ClusterListEntry

Create a value of ClusterListEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterState:ClusterListEntry', clusterListEntry_clusterState - The current state of this cluster. For information about the state of a specific node, see JobListEntry$JobState.

$sel:clusterId:ClusterListEntry', clusterListEntry_clusterId - The 39-character ID for the cluster that you want to list, for example CID123e4567-e89b-12d3-a456-426655440000.

$sel:creationDate:ClusterListEntry', clusterListEntry_creationDate - The creation date for this cluster.

$sel:description:ClusterListEntry', clusterListEntry_description - Defines an optional description of the cluster, for example Environmental Data Cluster-01.

clusterListEntry_clusterState :: Lens' ClusterListEntry (Maybe ClusterState)

The current state of this cluster. For information about the state of a specific node, see JobListEntry$JobState.

clusterListEntry_clusterId :: Lens' ClusterListEntry (Maybe Text)

The 39-character ID for the cluster that you want to list, for example CID123e4567-e89b-12d3-a456-426655440000.

clusterListEntry_creationDate :: Lens' ClusterListEntry (Maybe UTCTime)

The creation date for this cluster.

clusterListEntry_description :: Lens' ClusterListEntry (Maybe Text)

Defines an optional description of the cluster, for example Environmental Data Cluster-01.