Network.AWS.Snowball.Types.DataTransfer

Description

Documentation

data DataTransfer

Defines the real-time status of a Snow device's data transfer while the device is at AWS. This data is only available while a job has a JobState value of InProgress, for both import and export jobs.

See: newDataTransfer smart constructor.

Constructors

DataTransfer'

Fields

totalObjects :: Maybe Integer

The total number of objects for a transfer between a Snow device and Amazon S3. This value is set to 0 (zero) until all the keys that will be transferred have been listed.

bytesTransferred :: Maybe Integer

The number of bytes transferred between a Snow device and Amazon S3.

totalBytes :: Maybe Integer

The total bytes of data for a transfer between a Snow device and Amazon S3. This value is set to 0 (zero) until all the keys that will be transferred have been listed.

objectsTransferred :: Maybe Integer

The number of objects transferred between a Snow device and Amazon S3.

newDataTransfer :: DataTransfer

Create a value of DataTransfer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:totalObjects:DataTransfer', dataTransfer_totalObjects - The total number of objects for a transfer between a Snow device and Amazon S3. This value is set to 0 (zero) until all the keys that will be transferred have been listed.

$sel:bytesTransferred:DataTransfer', dataTransfer_bytesTransferred - The number of bytes transferred between a Snow device and Amazon S3.

$sel:totalBytes:DataTransfer', dataTransfer_totalBytes - The total bytes of data for a transfer between a Snow device and Amazon S3. This value is set to 0 (zero) until all the keys that will be transferred have been listed.

$sel:objectsTransferred:DataTransfer', dataTransfer_objectsTransferred - The number of objects transferred between a Snow device and Amazon S3.

dataTransfer_totalObjects :: Lens' DataTransfer (Maybe Integer)

The total number of objects for a transfer between a Snow device and Amazon S3. This value is set to 0 (zero) until all the keys that will be transferred have been listed.

dataTransfer_bytesTransferred :: Lens' DataTransfer (Maybe Integer)

The number of bytes transferred between a Snow device and Amazon S3.

dataTransfer_totalBytes :: Lens' DataTransfer (Maybe Integer)

The total bytes of data for a transfer between a Snow device and Amazon S3. This value is set to 0 (zero) until all the keys that will be transferred have been listed.

dataTransfer_objectsTransferred :: Lens' DataTransfer (Maybe Integer)

The number of objects transferred between a Snow device and Amazon S3.