Network.AWS.Snowball.Types.JobMetadata

Description

Documentation

data JobMetadata

Contains information about a specific job including shipping information, job status, and other important metadata. This information is returned as a part of the response syntax of the DescribeJob action.

See: newJobMetadata smart constructor.

Constructors

JobMetadata'

Fields

jobType :: Maybe JobType

The type of job.

kmsKeyARN :: Maybe Text

The Amazon Resource Name (ARN) for the AWS Key Management Service (AWS KMS) key associated with this job. This ARN was created using the CreateKey API action in AWS KMS.

remoteManagement :: Maybe RemoteManagement

Allows you to securely operate and manage Snowcone devices remotely from outside of your internal network. When set to INSTALLED_AUTOSTART, remote management will automatically be available when the device arrives at your location. Otherwise, you need to use the Snowball Client to manage the device.

jobId :: Maybe Text

The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000.

jobLogInfo :: Maybe JobLogs

Links to Amazon S3 presigned URLs for the job report and logs. For import jobs, the PDF job report becomes available at the end of the import process. For export jobs, your job report typically becomes available while the Snow device for your job part is being delivered to you.

notification :: Maybe Notification

The Amazon Simple Notification Service (Amazon SNS) notification settings associated with a specific job. The Notification object is returned as a part of the response syntax of the DescribeJob action in the JobMetadata data type.

jobState :: Maybe JobState

The current status of the jobs.

forwardingAddressId :: Maybe Text

The ID of the address that you want a job shipped to, after it will be shipped to its primary address. This field is not supported in most regions.

shippingDetails :: Maybe ShippingDetails

A job's shipping information, including inbound and outbound tracking numbers and shipping speed options.

addressId :: Maybe Text

The ID for the address that you want the Snow device shipped to.

snowballType :: Maybe SnowballType

The type of device used with this job.

dataTransferProgress :: Maybe DataTransfer

A value that defines the real-time status of a Snow device's data transfer while the device is at AWS. This data is only available while a job has a JobState value of InProgress, for both import and export jobs.

longTermPricingId :: Maybe Text

The ID of the long-term pricing type for the device.

resources :: Maybe JobResource

An array of S3Resource objects. Each S3Resource object represents an Amazon S3 bucket that your transferred data will be exported from or imported into.

onDeviceServiceConfiguration :: Maybe OnDeviceServiceConfiguration

Represents metadata and configuration settings for services on an AWS Snow Family device.

clusterId :: Maybe Text

The 39-character ID for the cluster, for example CID123e4567-e89b-12d3-a456-426655440000.

creationDate :: Maybe POSIX

The creation date for this job.

deviceConfiguration :: Maybe DeviceConfiguration

description :: Maybe Text

The description of the job, provided at job creation.

taxDocuments :: Maybe TaxDocuments

The metadata associated with the tax documents required in your AWS Region.

roleARN :: Maybe Text

The role ARN associated with this job. This ARN was created using the CreateRole API action in AWS Identity and Access Management (IAM).

snowballCapacityPreference :: Maybe SnowballCapacity

The Snow device capacity preference for this job, specified at job creation. In US regions, you can choose between 50 TB and 80 TB Snowballs. All other regions use 80 TB capacity Snowballs.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

newJobMetadata :: JobMetadata

Create a value of JobMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobType:JobMetadata', jobMetadata_jobType - The type of job.

$sel:kmsKeyARN:JobMetadata', jobMetadata_kmsKeyARN - The Amazon Resource Name (ARN) for the AWS Key Management Service (AWS KMS) key associated with this job. This ARN was created using the CreateKey API action in AWS KMS.

$sel:remoteManagement:JobMetadata', jobMetadata_remoteManagement - Allows you to securely operate and manage Snowcone devices remotely from outside of your internal network. When set to INSTALLED_AUTOSTART, remote management will automatically be available when the device arrives at your location. Otherwise, you need to use the Snowball Client to manage the device.

$sel:jobId:JobMetadata', jobMetadata_jobId - The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000.

$sel:jobLogInfo:JobMetadata', jobMetadata_jobLogInfo - Links to Amazon S3 presigned URLs for the job report and logs. For import jobs, the PDF job report becomes available at the end of the import process. For export jobs, your job report typically becomes available while the Snow device for your job part is being delivered to you.

$sel:notification:JobMetadata', jobMetadata_notification - The Amazon Simple Notification Service (Amazon SNS) notification settings associated with a specific job. The Notification object is returned as a part of the response syntax of the DescribeJob action in the JobMetadata data type.

$sel:jobState:JobMetadata', jobMetadata_jobState - The current status of the jobs.

$sel:forwardingAddressId:JobMetadata', jobMetadata_forwardingAddressId - The ID of the address that you want a job shipped to, after it will be shipped to its primary address. This field is not supported in most regions.

$sel:shippingDetails:JobMetadata', jobMetadata_shippingDetails - A job's shipping information, including inbound and outbound tracking numbers and shipping speed options.

$sel:addressId:JobMetadata', jobMetadata_addressId - The ID for the address that you want the Snow device shipped to.

$sel:snowballType:JobMetadata', jobMetadata_snowballType - The type of device used with this job.

$sel:dataTransferProgress:JobMetadata', jobMetadata_dataTransferProgress - A value that defines the real-time status of a Snow device's data transfer while the device is at AWS. This data is only available while a job has a JobState value of InProgress, for both import and export jobs.

$sel:longTermPricingId:JobMetadata', jobMetadata_longTermPricingId - The ID of the long-term pricing type for the device.

$sel:resources:JobMetadata', jobMetadata_resources - An array of S3Resource objects. Each S3Resource object represents an Amazon S3 bucket that your transferred data will be exported from or imported into.

$sel:onDeviceServiceConfiguration:JobMetadata', jobMetadata_onDeviceServiceConfiguration - Represents metadata and configuration settings for services on an AWS Snow Family device.

$sel:clusterId:JobMetadata', jobMetadata_clusterId - The 39-character ID for the cluster, for example CID123e4567-e89b-12d3-a456-426655440000.

$sel:creationDate:JobMetadata', jobMetadata_creationDate - The creation date for this job.

$sel:deviceConfiguration:JobMetadata', jobMetadata_deviceConfiguration - Undocumented member.

$sel:description:JobMetadata', jobMetadata_description - The description of the job, provided at job creation.

$sel:taxDocuments:JobMetadata', jobMetadata_taxDocuments - The metadata associated with the tax documents required in your AWS Region.

$sel:roleARN:JobMetadata', jobMetadata_roleARN - The role ARN associated with this job. This ARN was created using the CreateRole API action in AWS Identity and Access Management (IAM).

$sel:snowballCapacityPreference:JobMetadata', jobMetadata_snowballCapacityPreference - The Snow device capacity preference for this job, specified at job creation. In US regions, you can choose between 50 TB and 80 TB Snowballs. All other regions use 80 TB capacity Snowballs.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

jobMetadata_jobType :: Lens' JobMetadata (Maybe JobType)

The type of job.

jobMetadata_kmsKeyARN :: Lens' JobMetadata (Maybe Text)

The Amazon Resource Name (ARN) for the AWS Key Management Service (AWS KMS) key associated with this job. This ARN was created using the CreateKey API action in AWS KMS.

jobMetadata_remoteManagement :: Lens' JobMetadata (Maybe RemoteManagement)

Allows you to securely operate and manage Snowcone devices remotely from outside of your internal network. When set to INSTALLED_AUTOSTART, remote management will automatically be available when the device arrives at your location. Otherwise, you need to use the Snowball Client to manage the device.

jobMetadata_jobId :: Lens' JobMetadata (Maybe Text)

The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000.

jobMetadata_jobLogInfo :: Lens' JobMetadata (Maybe JobLogs)

Links to Amazon S3 presigned URLs for the job report and logs. For import jobs, the PDF job report becomes available at the end of the import process. For export jobs, your job report typically becomes available while the Snow device for your job part is being delivered to you.

jobMetadata_notification :: Lens' JobMetadata (Maybe Notification)

The Amazon Simple Notification Service (Amazon SNS) notification settings associated with a specific job. The Notification object is returned as a part of the response syntax of the DescribeJob action in the JobMetadata data type.

jobMetadata_jobState :: Lens' JobMetadata (Maybe JobState)

The current status of the jobs.

jobMetadata_forwardingAddressId :: Lens' JobMetadata (Maybe Text)

The ID of the address that you want a job shipped to, after it will be shipped to its primary address. This field is not supported in most regions.

jobMetadata_shippingDetails :: Lens' JobMetadata (Maybe ShippingDetails)

A job's shipping information, including inbound and outbound tracking numbers and shipping speed options.

jobMetadata_addressId :: Lens' JobMetadata (Maybe Text)

The ID for the address that you want the Snow device shipped to.

jobMetadata_snowballType :: Lens' JobMetadata (Maybe SnowballType)

The type of device used with this job.

jobMetadata_dataTransferProgress :: Lens' JobMetadata (Maybe DataTransfer)

A value that defines the real-time status of a Snow device's data transfer while the device is at AWS. This data is only available while a job has a JobState value of InProgress, for both import and export jobs.

jobMetadata_longTermPricingId :: Lens' JobMetadata (Maybe Text)

The ID of the long-term pricing type for the device.

jobMetadata_resources :: Lens' JobMetadata (Maybe JobResource)

An array of S3Resource objects. Each S3Resource object represents an Amazon S3 bucket that your transferred data will be exported from or imported into.

jobMetadata_onDeviceServiceConfiguration :: Lens' JobMetadata (Maybe OnDeviceServiceConfiguration)

Represents metadata and configuration settings for services on an AWS Snow Family device.

jobMetadata_clusterId :: Lens' JobMetadata (Maybe Text)

The 39-character ID for the cluster, for example CID123e4567-e89b-12d3-a456-426655440000.

jobMetadata_creationDate :: Lens' JobMetadata (Maybe UTCTime)

The creation date for this job.

jobMetadata_deviceConfiguration :: Lens' JobMetadata (Maybe DeviceConfiguration)

Undocumented member.

jobMetadata_description :: Lens' JobMetadata (Maybe Text)

The description of the job, provided at job creation.

jobMetadata_taxDocuments :: Lens' JobMetadata (Maybe TaxDocuments)

The metadata associated with the tax documents required in your AWS Region.

jobMetadata_roleARN :: Lens' JobMetadata (Maybe Text)

The role ARN associated with this job. This ARN was created using the CreateRole API action in AWS Identity and Access Management (IAM).

jobMetadata_snowballCapacityPreference :: Lens' JobMetadata (Maybe SnowballCapacity)

The Snow device capacity preference for this job, specified at job creation. In US regions, you can choose between 50 TB and 80 TB Snowballs. All other regions use 80 TB capacity Snowballs.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.