Network.AWS.Snowball.Types.Shipment

Description

Documentation

data Shipment

The Status and TrackingNumber information for an inbound or outbound shipment.

See: newShipment smart constructor.

Constructors

Shipment'

Fields

status :: Maybe Text

Status information for a shipment.

trackingNumber :: Maybe Text

The tracking number for this job. Using this tracking number with your region's carrier's website, you can track a Snow device as the carrier transports it.

For India, the carrier is Amazon Logistics. For all other regions, UPS is the carrier.

newShipment :: Shipment

Create a value of Shipment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Shipment', shipment_status - Status information for a shipment.

$sel:trackingNumber:Shipment', shipment_trackingNumber - The tracking number for this job. Using this tracking number with your region's carrier's website, you can track a Snow device as the carrier transports it.

For India, the carrier is Amazon Logistics. For all other regions, UPS is the carrier.

shipment_status :: Lens' Shipment (Maybe Text)

Status information for a shipment.

shipment_trackingNumber :: Lens' Shipment (Maybe Text)

The tracking number for this job. Using this tracking number with your region's carrier's website, you can track a Snow device as the carrier transports it.

For India, the carrier is Amazon Logistics. For all other regions, UPS is the carrier.