Network.AWS.Snowball.Types.ShippingDetails

Description

Documentation

data ShippingDetails

A job's shipping information, including inbound and outbound tracking numbers and shipping speed options.

See: newShippingDetails smart constructor.

Constructors

ShippingDetails'

Fields

shippingOption :: Maybe ShippingOption

The shipping speed for a particular job. This speed doesn't dictate how soon you'll get the Snow device from the job's creation date. This speed represents how quickly it moves to its destination while in transit. Regional shipping speeds are as follows:

In Australia, you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snow devices are delivered in one to seven days.

In the United States of America (US), you have access to one-day shipping and two-day shipping.

outboundShipment :: Maybe Shipment

The Status and TrackingNumber values for a Snow device being delivered to the address that you specified for a particular job.

inboundShipment :: Maybe Shipment

The Status and TrackingNumber values for a Snow device being returned to AWS for a particular job.

newShippingDetails :: ShippingDetails

Create a value of ShippingDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shippingOption:ShippingDetails', shippingDetails_shippingOption - The shipping speed for a particular job. This speed doesn't dictate how soon you'll get the Snow device from the job's creation date. This speed represents how quickly it moves to its destination while in transit. Regional shipping speeds are as follows:

In Australia, you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snow devices are delivered in one to seven days.

In the United States of America (US), you have access to one-day shipping and two-day shipping.

$sel:outboundShipment:ShippingDetails', shippingDetails_outboundShipment - The Status and TrackingNumber values for a Snow device being delivered to the address that you specified for a particular job.

$sel:inboundShipment:ShippingDetails', shippingDetails_inboundShipment - The Status and TrackingNumber values for a Snow device being returned to AWS for a particular job.

shippingDetails_shippingOption :: Lens' ShippingDetails (Maybe ShippingOption)

The shipping speed for a particular job. This speed doesn't dictate how soon you'll get the Snow device from the job's creation date. This speed represents how quickly it moves to its destination while in transit. Regional shipping speeds are as follows:

In Australia, you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snow devices are delivered in one to seven days.

In the United States of America (US), you have access to one-day shipping and two-day shipping.

shippingDetails_outboundShipment :: Lens' ShippingDetails (Maybe Shipment)

The Status and TrackingNumber values for a Snow device being delivered to the address that you specified for a particular job.

shippingDetails_inboundShipment :: Lens' ShippingDetails (Maybe Shipment)

The Status and TrackingNumber values for a Snow device being returned to AWS for a particular job.