Network.AWS.Snowball.UpdateCluster

Description

While a cluster's ClusterState value is in the AwaitingQuorum state, you can update some of the information associated with a cluster. Once the cluster changes to a different job state, usually 60 minutes after the cluster being created, this action is no longer available.

Creating a Request

data UpdateCluster

See: newUpdateCluster smart constructor.

Constructors

UpdateCluster'

Fields

roleARN :: Maybe Text

The new role Amazon Resource Name (ARN) that you want to associate with this cluster. To create a role ARN, use the CreateRole API action in AWS Identity and Access Management (IAM).

shippingOption :: Maybe ShippingOption

The updated shipping option value of this cluster's ShippingDetails object.

onDeviceServiceConfiguration :: Maybe OnDeviceServiceConfiguration

Specifies the service or services on the Snow Family device that your transferred data will be exported from or imported into. AWS Snow Family supports Amazon S3 and NFS (Network File System).

resources :: Maybe JobResource

The updated arrays of JobResource objects that can include updated S3Resource objects or LambdaResource objects.

description :: Maybe Text

The updated description of this cluster.

addressId :: Maybe Text

The ID of the updated Address object.

forwardingAddressId :: Maybe Text

The updated ID for the forwarding address for a cluster. This field is not supported in most regions.

notification :: Maybe Notification

The new or updated Notification object.

clusterId :: Text

The cluster ID of the cluster that you want to update, for example CID123e4567-e89b-12d3-a456-426655440000.

newUpdateCluster

Arguments

:: Text
$sel:clusterId:UpdateCluster'

-> UpdateCluster

Create a value of UpdateCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleARN:UpdateCluster', updateCluster_roleARN - The new role Amazon Resource Name (ARN) that you want to associate with this cluster. To create a role ARN, use the CreateRole API action in AWS Identity and Access Management (IAM).

$sel:shippingOption:UpdateCluster', updateCluster_shippingOption - The updated shipping option value of this cluster's ShippingDetails object.

$sel:onDeviceServiceConfiguration:UpdateCluster', updateCluster_onDeviceServiceConfiguration - Specifies the service or services on the Snow Family device that your transferred data will be exported from or imported into. AWS Snow Family supports Amazon S3 and NFS (Network File System).

$sel:resources:UpdateCluster', updateCluster_resources - The updated arrays of JobResource objects that can include updated S3Resource objects or LambdaResource objects.

$sel:description:UpdateCluster', updateCluster_description - The updated description of this cluster.

$sel:addressId:UpdateCluster', updateCluster_addressId - The ID of the updated Address object.

$sel:forwardingAddressId:UpdateCluster', updateCluster_forwardingAddressId - The updated ID for the forwarding address for a cluster. This field is not supported in most regions.

$sel:notification:UpdateCluster', updateCluster_notification - The new or updated Notification object.

$sel:clusterId:UpdateCluster', updateCluster_clusterId - The cluster ID of the cluster that you want to update, for example CID123e4567-e89b-12d3-a456-426655440000.

Request Lenses

updateCluster_roleARN :: Lens' UpdateCluster (Maybe Text)

The new role Amazon Resource Name (ARN) that you want to associate with this cluster. To create a role ARN, use the CreateRole API action in AWS Identity and Access Management (IAM).

updateCluster_shippingOption :: Lens' UpdateCluster (Maybe ShippingOption)

The updated shipping option value of this cluster's ShippingDetails object.

updateCluster_onDeviceServiceConfiguration :: Lens' UpdateCluster (Maybe OnDeviceServiceConfiguration)

Specifies the service or services on the Snow Family device that your transferred data will be exported from or imported into. AWS Snow Family supports Amazon S3 and NFS (Network File System).

updateCluster_resources :: Lens' UpdateCluster (Maybe JobResource)

The updated arrays of JobResource objects that can include updated S3Resource objects or LambdaResource objects.

updateCluster_description :: Lens' UpdateCluster (Maybe Text)

The updated description of this cluster.

updateCluster_addressId :: Lens' UpdateCluster (Maybe Text)

The ID of the updated Address object.

updateCluster_forwardingAddressId :: Lens' UpdateCluster (Maybe Text)

The updated ID for the forwarding address for a cluster. This field is not supported in most regions.

updateCluster_notification :: Lens' UpdateCluster (Maybe Notification)

The new or updated Notification object.

updateCluster_clusterId :: Lens' UpdateCluster Text

The cluster ID of the cluster that you want to update, for example CID123e4567-e89b-12d3-a456-426655440000.

Destructuring the Response

data UpdateClusterResponse

See: newUpdateClusterResponse smart constructor.

Constructors

UpdateClusterResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateClusterResponse

Arguments

:: Int
$sel:httpStatus:UpdateClusterResponse'

-> UpdateClusterResponse

Create a value of UpdateClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateClusterResponse', updateClusterResponse_httpStatus - The response's http status code.

Response Lenses

updateClusterResponse_httpStatus :: Lens' UpdateClusterResponse Int

The response's http status code.