Network.AWS.Snowball.UpdateJob

Description

While a job's JobState value is New, you can update some of the information associated with a job. Once the job changes to a different job state, usually within 60 minutes of the job being created, this action is no longer available.

Creating a Request

data UpdateJob

See: newUpdateJob smart constructor.

Constructors

UpdateJob'

Fields

roleARN :: Maybe Text

The new role Amazon Resource Name (ARN) that you want to associate with this job. To create a role ARN, use the CreateRoleAWS Identity and Access Management (IAM) API action.

shippingOption :: Maybe ShippingOption

The updated shipping option value of this job's ShippingDetails object.

onDeviceServiceConfiguration :: Maybe OnDeviceServiceConfiguration

Specifies the service or services on the Snow Family device that your transferred data will be exported from or imported into. AWS Snow Family supports Amazon S3 and NFS (Network File System).

snowballCapacityPreference :: Maybe SnowballCapacity

The updated SnowballCapacityPreference of this job's JobMetadata object. The 50 TB Snowballs are only available in the US regions.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

resources :: Maybe JobResource

The updated JobResource object, or the updated JobResource object.

description :: Maybe Text

The updated description of this job's JobMetadata object.

addressId :: Maybe Text

The ID of the updated Address object.

forwardingAddressId :: Maybe Text

The updated ID for the forwarding address for a job. This field is not supported in most regions.

notification :: Maybe Notification

The new or updated Notification object.

jobId :: Text

The job ID of the job that you want to update, for example JID123e4567-e89b-12d3-a456-426655440000.

newUpdateJob

Arguments

:: Text
$sel:jobId:UpdateJob'

-> UpdateJob

Create a value of UpdateJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleARN:UpdateJob', updateJob_roleARN - The new role Amazon Resource Name (ARN) that you want to associate with this job. To create a role ARN, use the CreateRoleAWS Identity and Access Management (IAM) API action.

$sel:shippingOption:UpdateJob', updateJob_shippingOption - The updated shipping option value of this job's ShippingDetails object.

$sel:onDeviceServiceConfiguration:UpdateJob', updateJob_onDeviceServiceConfiguration - Specifies the service or services on the Snow Family device that your transferred data will be exported from or imported into. AWS Snow Family supports Amazon S3 and NFS (Network File System).

$sel:snowballCapacityPreference:UpdateJob', updateJob_snowballCapacityPreference - The updated SnowballCapacityPreference of this job's JobMetadata object. The 50 TB Snowballs are only available in the US regions.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

$sel:resources:UpdateJob', updateJob_resources - The updated JobResource object, or the updated JobResource object.

$sel:description:UpdateJob', updateJob_description - The updated description of this job's JobMetadata object.

$sel:addressId:UpdateJob', updateJob_addressId - The ID of the updated Address object.

$sel:forwardingAddressId:UpdateJob', updateJob_forwardingAddressId - The updated ID for the forwarding address for a job. This field is not supported in most regions.

$sel:notification:UpdateJob', updateJob_notification - The new or updated Notification object.

$sel:jobId:UpdateJob', updateJob_jobId - The job ID of the job that you want to update, for example JID123e4567-e89b-12d3-a456-426655440000.

Request Lenses

updateJob_roleARN :: Lens' UpdateJob (Maybe Text)

The new role Amazon Resource Name (ARN) that you want to associate with this job. To create a role ARN, use the CreateRoleAWS Identity and Access Management (IAM) API action.

updateJob_shippingOption :: Lens' UpdateJob (Maybe ShippingOption)

The updated shipping option value of this job's ShippingDetails object.

updateJob_onDeviceServiceConfiguration :: Lens' UpdateJob (Maybe OnDeviceServiceConfiguration)

Specifies the service or services on the Snow Family device that your transferred data will be exported from or imported into. AWS Snow Family supports Amazon S3 and NFS (Network File System).

updateJob_snowballCapacityPreference :: Lens' UpdateJob (Maybe SnowballCapacity)

The updated SnowballCapacityPreference of this job's JobMetadata object. The 50 TB Snowballs are only available in the US regions.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

updateJob_resources :: Lens' UpdateJob (Maybe JobResource)

The updated JobResource object, or the updated JobResource object.

updateJob_description :: Lens' UpdateJob (Maybe Text)

The updated description of this job's JobMetadata object.

updateJob_addressId :: Lens' UpdateJob (Maybe Text)

The ID of the updated Address object.

updateJob_forwardingAddressId :: Lens' UpdateJob (Maybe Text)

The updated ID for the forwarding address for a job. This field is not supported in most regions.

updateJob_notification :: Lens' UpdateJob (Maybe Notification)

The new or updated Notification object.

updateJob_jobId :: Lens' UpdateJob Text

The job ID of the job that you want to update, for example JID123e4567-e89b-12d3-a456-426655440000.

Destructuring the Response

data UpdateJobResponse

See: newUpdateJobResponse smart constructor.

Constructors

UpdateJobResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateJobResponse

Arguments

:: Int
$sel:httpStatus:UpdateJobResponse'

-> UpdateJobResponse

Create a value of UpdateJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateJobResponse', updateJobResponse_httpStatus - The response's http status code.

Response Lenses

updateJobResponse_httpStatus :: Lens' UpdateJobResponse Int

The response's http status code.