Network.AWS.StepFunctions.CreateActivity

Description

Creates an activity. An activity is a task that you write in any programming language and host on any machine that has access to AWS Step Functions. Activities must poll Step Functions using the GetActivityTask API action and respond using SendTask* API actions. This function lets Step Functions know the existence of your activity and returns an identifier for use in a state machine and when polling from the activity.

This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

CreateActivity is an idempotent API. Subsequent requests won’t create a duplicate resource if it was already created. CreateActivity's idempotency check is based on the activity name. If a following request has different tags values, Step Functions will ignore these differences and treat it as an idempotent request of the previous. In this case, tags will not be updated, even if they are different.

Creating a Request

data CreateActivity

See: newCreateActivity smart constructor.

Constructors

CreateActivity'

Fields

tags :: Maybe [Tag]

The list of tags to add to a resource.

An array of key-value pairs. For more information, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide, and Controlling Access Using IAM Tags.

Tags may only contain Unicode letters, digits, white space, or these symbols: _ . : / = + - @.

name :: Text

The name of the activity to create. This name must be unique for your AWS account and region for 90 days. For more information, see Limits Related to State Machine Executions in the AWS Step Functions Developer Guide.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

newCreateActivity

Arguments

:: Text
$sel:name:CreateActivity'

-> CreateActivity

Create a value of CreateActivity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateActivity', createActivity_tags - The list of tags to add to a resource.

An array of key-value pairs. For more information, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide, and Controlling Access Using IAM Tags.

Tags may only contain Unicode letters, digits, white space, or these symbols: _ . : / = + - @.

$sel:name:CreateActivity', createActivity_name - The name of the activity to create. This name must be unique for your AWS account and region for 90 days. For more information, see Limits Related to State Machine Executions in the AWS Step Functions Developer Guide.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

Request Lenses

createActivity_tags :: Lens' CreateActivity (Maybe [Tag])

The list of tags to add to a resource.

An array of key-value pairs. For more information, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide, and Controlling Access Using IAM Tags.

Tags may only contain Unicode letters, digits, white space, or these symbols: _ . : / = + - @.

createActivity_name :: Lens' CreateActivity Text

The name of the activity to create. This name must be unique for your AWS account and region for 90 days. For more information, see Limits Related to State Machine Executions in the AWS Step Functions Developer Guide.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

Destructuring the Response

data CreateActivityResponse

See: newCreateActivityResponse smart constructor.

Constructors

CreateActivityResponse'

Fields

httpStatus :: Int

The response's http status code.

activityArn :: Text

The Amazon Resource Name (ARN) that identifies the created activity.

creationDate :: POSIX

The date the activity is created.

newCreateActivityResponse

Arguments

:: Int
$sel:httpStatus:CreateActivityResponse'

-> Text
$sel:activityArn:CreateActivityResponse'

-> UTCTime
$sel:creationDate:CreateActivityResponse'

-> CreateActivityResponse

Create a value of CreateActivityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateActivityResponse', createActivityResponse_httpStatus - The response's http status code.

$sel:activityArn:CreateActivityResponse', createActivityResponse_activityArn - The Amazon Resource Name (ARN) that identifies the created activity.

$sel:creationDate:CreateActivityResponse', createActivityResponse_creationDate - The date the activity is created.

Response Lenses

createActivityResponse_httpStatus :: Lens' CreateActivityResponse Int

The response's http status code.

createActivityResponse_activityArn :: Lens' CreateActivityResponse Text

The Amazon Resource Name (ARN) that identifies the created activity.

createActivityResponse_creationDate :: Lens' CreateActivityResponse UTCTime

The date the activity is created.