Network.AWS.StepFunctions.DescribeExecution

Description

Describes an execution.

This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

This API action is not supported by EXPRESS state machines.

Creating a Request

data DescribeExecution

See: newDescribeExecution smart constructor.

Constructors

DescribeExecution'

Fields

executionArn :: Text

The Amazon Resource Name (ARN) of the execution to describe.

newDescribeExecution

Arguments

:: Text
$sel:executionArn:DescribeExecution'

-> DescribeExecution

Create a value of DescribeExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:executionArn:DescribeExecution', describeExecution_executionArn - The Amazon Resource Name (ARN) of the execution to describe.

Request Lenses

describeExecution_executionArn :: Lens' DescribeExecution Text

The Amazon Resource Name (ARN) of the execution to describe.

Destructuring the Response

data DescribeExecutionResponse

See: newDescribeExecutionResponse smart constructor.

Constructors

DescribeExecutionResponse'

Fields

stopDate :: Maybe POSIX

If the execution has already ended, the date the execution stopped.

inputDetails :: Maybe CloudWatchEventsExecutionDataDetails

input :: Maybe (Sensitive Text)

The string that contains the JSON input data of the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

output :: Maybe (Sensitive Text)

The JSON output data of the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

This field is set only if the execution succeeds. If the execution fails, this field is null.

name :: Maybe Text

The name of the execution.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

traceHeader :: Maybe Text

The AWS X-Ray trace header that was passed to the execution.

outputDetails :: Maybe CloudWatchEventsExecutionDataDetails

httpStatus :: Int

The response's http status code.

executionArn :: Text

The Amazon Resource Name (ARN) that identifies the execution.

stateMachineArn :: Text

The Amazon Resource Name (ARN) of the executed stated machine.

status :: ExecutionStatus

The current status of the execution.

startDate :: POSIX

The date the execution is started.

newDescribeExecutionResponse

Arguments

:: Int
$sel:httpStatus:DescribeExecutionResponse'

-> Text
$sel:executionArn:DescribeExecution'

-> Text
$sel:stateMachineArn:DescribeExecutionResponse'

-> ExecutionStatus
$sel:status:DescribeExecutionResponse'

-> UTCTime
$sel:startDate:DescribeExecutionResponse'

-> DescribeExecutionResponse

Create a value of DescribeExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stopDate:DescribeExecutionResponse', describeExecutionResponse_stopDate - If the execution has already ended, the date the execution stopped.

$sel:inputDetails:DescribeExecutionResponse', describeExecutionResponse_inputDetails - Undocumented member.

$sel:input:DescribeExecutionResponse', describeExecutionResponse_input - The string that contains the JSON input data of the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

$sel:output:DescribeExecutionResponse', describeExecutionResponse_output - The JSON output data of the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

This field is set only if the execution succeeds. If the execution fails, this field is null.

$sel:name:DescribeExecutionResponse', describeExecutionResponse_name - The name of the execution.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

$sel:traceHeader:DescribeExecutionResponse', describeExecutionResponse_traceHeader - The AWS X-Ray trace header that was passed to the execution.

$sel:outputDetails:DescribeExecutionResponse', describeExecutionResponse_outputDetails - Undocumented member.

$sel:httpStatus:DescribeExecutionResponse', describeExecutionResponse_httpStatus - The response's http status code.

$sel:executionArn:DescribeExecution', describeExecutionResponse_executionArn - The Amazon Resource Name (ARN) that identifies the execution.

$sel:stateMachineArn:DescribeExecutionResponse', describeExecutionResponse_stateMachineArn - The Amazon Resource Name (ARN) of the executed stated machine.

$sel:status:DescribeExecutionResponse', describeExecutionResponse_status - The current status of the execution.

$sel:startDate:DescribeExecutionResponse', describeExecutionResponse_startDate - The date the execution is started.

Response Lenses

describeExecutionResponse_stopDate :: Lens' DescribeExecutionResponse (Maybe UTCTime)

If the execution has already ended, the date the execution stopped.

describeExecutionResponse_inputDetails :: Lens' DescribeExecutionResponse (Maybe CloudWatchEventsExecutionDataDetails)

Undocumented member.

describeExecutionResponse_input :: Lens' DescribeExecutionResponse (Maybe Text)

The string that contains the JSON input data of the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

describeExecutionResponse_output :: Lens' DescribeExecutionResponse (Maybe Text)

The JSON output data of the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

This field is set only if the execution succeeds. If the execution fails, this field is null.

describeExecutionResponse_name :: Lens' DescribeExecutionResponse (Maybe Text)

The name of the execution.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

describeExecutionResponse_traceHeader :: Lens' DescribeExecutionResponse (Maybe Text)

The AWS X-Ray trace header that was passed to the execution.

describeExecutionResponse_outputDetails :: Lens' DescribeExecutionResponse (Maybe CloudWatchEventsExecutionDataDetails)

Undocumented member.

describeExecutionResponse_httpStatus :: Lens' DescribeExecutionResponse Int

The response's http status code.

describeExecutionResponse_executionArn :: Lens' DescribeExecutionResponse Text

The Amazon Resource Name (ARN) that identifies the execution.

describeExecutionResponse_stateMachineArn :: Lens' DescribeExecutionResponse Text

The Amazon Resource Name (ARN) of the executed stated machine.

describeExecutionResponse_status :: Lens' DescribeExecutionResponse ExecutionStatus

The current status of the execution.

describeExecutionResponse_startDate :: Lens' DescribeExecutionResponse UTCTime

The date the execution is started.