Network.AWS.StepFunctions.DescribeStateMachine

Description

Describes a state machine.

This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

Creating a Request

data DescribeStateMachine

See: newDescribeStateMachine smart constructor.

Constructors

DescribeStateMachine'

Fields

stateMachineArn :: Text

The Amazon Resource Name (ARN) of the state machine to describe.

newDescribeStateMachine

Arguments

:: Text
$sel:stateMachineArn:DescribeStateMachine'

-> DescribeStateMachine

Create a value of DescribeStateMachine with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stateMachineArn:DescribeStateMachine', describeStateMachine_stateMachineArn - The Amazon Resource Name (ARN) of the state machine to describe.

Request Lenses

describeStateMachine_stateMachineArn :: Lens' DescribeStateMachine Text

The Amazon Resource Name (ARN) of the state machine to describe.

Destructuring the Response

data DescribeStateMachineResponse

See: newDescribeStateMachineResponse smart constructor.

Constructors

DescribeStateMachineResponse'

Fields

status :: Maybe StateMachineStatus

The current status of the state machine.

tracingConfiguration :: Maybe TracingConfiguration

Selects whether AWS X-Ray tracing is enabled.

loggingConfiguration :: Maybe LoggingConfiguration

httpStatus :: Int

The response's http status code.

stateMachineArn :: Text

The Amazon Resource Name (ARN) that identifies the state machine.

name :: Text

The name of the state machine.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

definition :: Sensitive Text

The Amazon States Language definition of the state machine. See Amazon States Language.

roleArn :: Text

The Amazon Resource Name (ARN) of the IAM role used when creating this state machine. (The IAM role maintains security by granting Step Functions access to AWS resources.)

type' :: StateMachineType

The type of the state machine (STANDARD or EXPRESS).

creationDate :: POSIX

The date the state machine is created.

newDescribeStateMachineResponse

Arguments

:: Int
$sel:httpStatus:DescribeStateMachineResponse'

-> Text
$sel:stateMachineArn:DescribeStateMachine'

-> Text
$sel:name:DescribeStateMachineResponse'

-> Text
$sel:definition:DescribeStateMachineResponse'

-> Text
$sel:roleArn:DescribeStateMachineResponse'

-> StateMachineType
$sel:type':DescribeStateMachineResponse'

-> UTCTime
$sel:creationDate:DescribeStateMachineResponse'

-> DescribeStateMachineResponse

Create a value of DescribeStateMachineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeStateMachineResponse', describeStateMachineResponse_status - The current status of the state machine.

$sel:tracingConfiguration:DescribeStateMachineResponse', describeStateMachineResponse_tracingConfiguration - Selects whether AWS X-Ray tracing is enabled.

$sel:loggingConfiguration:DescribeStateMachineResponse', describeStateMachineResponse_loggingConfiguration - Undocumented member.

$sel:httpStatus:DescribeStateMachineResponse', describeStateMachineResponse_httpStatus - The response's http status code.

$sel:stateMachineArn:DescribeStateMachine', describeStateMachineResponse_stateMachineArn - The Amazon Resource Name (ARN) that identifies the state machine.

$sel:name:DescribeStateMachineResponse', describeStateMachineResponse_name - The name of the state machine.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

$sel:definition:DescribeStateMachineResponse', describeStateMachineResponse_definition - The Amazon States Language definition of the state machine. See Amazon States Language.

$sel:roleArn:DescribeStateMachineResponse', describeStateMachineResponse_roleArn - The Amazon Resource Name (ARN) of the IAM role used when creating this state machine. (The IAM role maintains security by granting Step Functions access to AWS resources.)

$sel:type':DescribeStateMachineResponse', describeStateMachineResponse_type - The type of the state machine (STANDARD or EXPRESS).

$sel:creationDate:DescribeStateMachineResponse', describeStateMachineResponse_creationDate - The date the state machine is created.

Response Lenses

describeStateMachineResponse_status :: Lens' DescribeStateMachineResponse (Maybe StateMachineStatus)

The current status of the state machine.

describeStateMachineResponse_tracingConfiguration :: Lens' DescribeStateMachineResponse (Maybe TracingConfiguration)

Selects whether AWS X-Ray tracing is enabled.

describeStateMachineResponse_loggingConfiguration :: Lens' DescribeStateMachineResponse (Maybe LoggingConfiguration)

Undocumented member.

describeStateMachineResponse_httpStatus :: Lens' DescribeStateMachineResponse Int

The response's http status code.

describeStateMachineResponse_stateMachineArn :: Lens' DescribeStateMachineResponse Text

The Amazon Resource Name (ARN) that identifies the state machine.

describeStateMachineResponse_name :: Lens' DescribeStateMachineResponse Text

The name of the state machine.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

describeStateMachineResponse_definition :: Lens' DescribeStateMachineResponse Text

The Amazon States Language definition of the state machine. See Amazon States Language.

describeStateMachineResponse_roleArn :: Lens' DescribeStateMachineResponse Text

The Amazon Resource Name (ARN) of the IAM role used when creating this state machine. (The IAM role maintains security by granting Step Functions access to AWS resources.)

describeStateMachineResponse_type :: Lens' DescribeStateMachineResponse StateMachineType

The type of the state machine (STANDARD or EXPRESS).

describeStateMachineResponse_creationDate :: Lens' DescribeStateMachineResponse UTCTime

The date the state machine is created.