Network.AWS.StepFunctions.GetActivityTask

Description

Used by workers to retrieve a task (with the specified activity ARN) which has been scheduled for execution by a running state machine. This initiates a long poll, where the service holds the HTTP connection open and responds as soon as a task becomes available (i.e. an execution of a task of this type is needed.) The maximum time the service holds on to the request before responding is 60 seconds. If no task is available within 60 seconds, the poll returns a taskToken with a null string.

Workers should set their client side socket timeout to at least 65 seconds (5 seconds higher than the maximum time the service may hold the poll request).

Polling with GetActivityTask can cause latency in some implementations. See Avoid Latency When Polling for Activity Tasks in the Step Functions Developer Guide.

Creating a Request

data GetActivityTask

See: newGetActivityTask smart constructor.

Constructors

GetActivityTask'

Fields

workerName :: Maybe Text

You can provide an arbitrary name in order to identify the worker that the task is assigned to. This name is used when it is logged in the execution history.

activityArn :: Text

The Amazon Resource Name (ARN) of the activity to retrieve tasks from (assigned when you create the task using CreateActivity.)

newGetActivityTask

Arguments

:: Text
$sel:activityArn:GetActivityTask'

-> GetActivityTask

Create a value of GetActivityTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workerName:GetActivityTask', getActivityTask_workerName - You can provide an arbitrary name in order to identify the worker that the task is assigned to. This name is used when it is logged in the execution history.

$sel:activityArn:GetActivityTask', getActivityTask_activityArn - The Amazon Resource Name (ARN) of the activity to retrieve tasks from (assigned when you create the task using CreateActivity.)

Request Lenses

getActivityTask_workerName :: Lens' GetActivityTask (Maybe Text)

You can provide an arbitrary name in order to identify the worker that the task is assigned to. This name is used when it is logged in the execution history.

getActivityTask_activityArn :: Lens' GetActivityTask Text

The Amazon Resource Name (ARN) of the activity to retrieve tasks from (assigned when you create the task using CreateActivity.)

Destructuring the Response

data GetActivityTaskResponse

See: newGetActivityTaskResponse smart constructor.

Constructors

GetActivityTaskResponse'

Fields

input :: Maybe (Sensitive Text)

The string that contains the JSON input data for the task. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

taskToken :: Maybe Text

A token that identifies the scheduled task. This token must be copied and included in subsequent calls to SendTaskHeartbeat, SendTaskSuccess or SendTaskFailure in order to report the progress or completion of the task.

httpStatus :: Int

The response's http status code.

newGetActivityTaskResponse

Arguments

:: Int
$sel:httpStatus:GetActivityTaskResponse'

-> GetActivityTaskResponse

Create a value of GetActivityTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:input:GetActivityTaskResponse', getActivityTaskResponse_input - The string that contains the JSON input data for the task. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

$sel:taskToken:GetActivityTaskResponse', getActivityTaskResponse_taskToken - A token that identifies the scheduled task. This token must be copied and included in subsequent calls to SendTaskHeartbeat, SendTaskSuccess or SendTaskFailure in order to report the progress or completion of the task.

$sel:httpStatus:GetActivityTaskResponse', getActivityTaskResponse_httpStatus - The response's http status code.

Response Lenses

getActivityTaskResponse_input :: Lens' GetActivityTaskResponse (Maybe Text)

The string that contains the JSON input data for the task. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

getActivityTaskResponse_taskToken :: Lens' GetActivityTaskResponse (Maybe Text)

A token that identifies the scheduled task. This token must be copied and included in subsequent calls to SendTaskHeartbeat, SendTaskSuccess or SendTaskFailure in order to report the progress or completion of the task.

getActivityTaskResponse_httpStatus :: Lens' GetActivityTaskResponse Int

The response's http status code.