Network.AWS.StepFunctions.Lens

Description

Operations

DeleteActivity

deleteActivity_activityArn :: Lens' DeleteActivity Text

The Amazon Resource Name (ARN) of the activity to delete.

deleteActivityResponse_httpStatus :: Lens' DeleteActivityResponse Int

The response's http status code.

DescribeStateMachine

describeStateMachine_stateMachineArn :: Lens' DescribeStateMachine Text

The Amazon Resource Name (ARN) of the state machine to describe.

describeStateMachineResponse_status :: Lens' DescribeStateMachineResponse (Maybe StateMachineStatus)

The current status of the state machine.

describeStateMachineResponse_tracingConfiguration :: Lens' DescribeStateMachineResponse (Maybe TracingConfiguration)

Selects whether AWS X-Ray tracing is enabled.

describeStateMachineResponse_loggingConfiguration :: Lens' DescribeStateMachineResponse (Maybe LoggingConfiguration)

Undocumented member.

describeStateMachineResponse_httpStatus :: Lens' DescribeStateMachineResponse Int

The response's http status code.

describeStateMachineResponse_stateMachineArn :: Lens' DescribeStateMachineResponse Text

The Amazon Resource Name (ARN) that identifies the state machine.

describeStateMachineResponse_name :: Lens' DescribeStateMachineResponse Text

The name of the state machine.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

describeStateMachineResponse_definition :: Lens' DescribeStateMachineResponse Text

The Amazon States Language definition of the state machine. See Amazon States Language.

describeStateMachineResponse_roleArn :: Lens' DescribeStateMachineResponse Text

The Amazon Resource Name (ARN) of the IAM role used when creating this state machine. (The IAM role maintains security by granting Step Functions access to AWS resources.)

describeStateMachineResponse_type :: Lens' DescribeStateMachineResponse StateMachineType

The type of the state machine (STANDARD or EXPRESS).

describeStateMachineResponse_creationDate :: Lens' DescribeStateMachineResponse UTCTime

The date the state machine is created.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) for the Step Functions state machine or activity.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

An array of tags associated with the resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

StopExecution

stopExecution_error :: Lens' StopExecution (Maybe Text)

The error code of the failure.

stopExecution_cause :: Lens' StopExecution (Maybe Text)

A more detailed explanation of the cause of the failure.

stopExecution_executionArn :: Lens' StopExecution Text

The Amazon Resource Name (ARN) of the execution to stop.

stopExecutionResponse_httpStatus :: Lens' StopExecutionResponse Int

The response's http status code.

stopExecutionResponse_stopDate :: Lens' StopExecutionResponse UTCTime

The date the execution is stopped.

DescribeActivity

describeActivity_activityArn :: Lens' DescribeActivity Text

The Amazon Resource Name (ARN) of the activity to describe.

describeActivityResponse_httpStatus :: Lens' DescribeActivityResponse Int

The response's http status code.

describeActivityResponse_activityArn :: Lens' DescribeActivityResponse Text

The Amazon Resource Name (ARN) that identifies the activity.

describeActivityResponse_name :: Lens' DescribeActivityResponse Text

The name of the activity.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

describeActivityResponse_creationDate :: Lens' DescribeActivityResponse UTCTime

The date the activity is created.

ListStateMachines

listStateMachines_nextToken :: Lens' ListStateMachines (Maybe Text)

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

listStateMachines_maxResults :: Lens' ListStateMachines (Maybe Natural)

The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000. A value of 0 uses the default.

This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.

listStateMachinesResponse_nextToken :: Lens' ListStateMachinesResponse (Maybe Text)

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

listStateMachinesResponse_httpStatus :: Lens' ListStateMachinesResponse Int

The response's http status code.

listStateMachinesResponse_stateMachines :: Lens' ListStateMachinesResponse [StateMachineListItem]

Undocumented member.

ListExecutions

listExecutions_statusFilter :: Lens' ListExecutions (Maybe ExecutionStatus)

If specified, only list the executions whose current execution status matches the given filter.

listExecutions_nextToken :: Lens' ListExecutions (Maybe Text)

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

listExecutions_maxResults :: Lens' ListExecutions (Maybe Natural)

The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000. A value of 0 uses the default.

This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.

listExecutions_stateMachineArn :: Lens' ListExecutions Text

The Amazon Resource Name (ARN) of the state machine whose executions is listed.

listExecutionsResponse_nextToken :: Lens' ListExecutionsResponse (Maybe Text)

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

listExecutionsResponse_httpStatus :: Lens' ListExecutionsResponse Int

The response's http status code.

listExecutionsResponse_executions :: Lens' ListExecutionsResponse [ExecutionListItem]

The list of matching executions.

DeleteStateMachine

deleteStateMachine_stateMachineArn :: Lens' DeleteStateMachine Text

The Amazon Resource Name (ARN) of the state machine to delete.

deleteStateMachineResponse_httpStatus :: Lens' DeleteStateMachineResponse Int

The response's http status code.

UpdateStateMachine

updateStateMachine_definition :: Lens' UpdateStateMachine (Maybe Text)

The Amazon States Language definition of the state machine. See Amazon States Language.

updateStateMachine_tracingConfiguration :: Lens' UpdateStateMachine (Maybe TracingConfiguration)

Selects whether AWS X-Ray tracing is enabled.

updateStateMachine_loggingConfiguration :: Lens' UpdateStateMachine (Maybe LoggingConfiguration)

The LoggingConfiguration data type is used to set CloudWatch Logs options.

updateStateMachine_roleArn :: Lens' UpdateStateMachine (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role of the state machine.

updateStateMachine_stateMachineArn :: Lens' UpdateStateMachine Text

The Amazon Resource Name (ARN) of the state machine.

updateStateMachineResponse_httpStatus :: Lens' UpdateStateMachineResponse Int

The response's http status code.

updateStateMachineResponse_updateDate :: Lens' UpdateStateMachineResponse UTCTime

The date and time the state machine was updated.

DescribeStateMachineForExecution

describeStateMachineForExecution_executionArn :: Lens' DescribeStateMachineForExecution Text

The Amazon Resource Name (ARN) of the execution you want state machine information for.

describeStateMachineForExecutionResponse_tracingConfiguration :: Lens' DescribeStateMachineForExecutionResponse (Maybe TracingConfiguration)

Selects whether AWS X-Ray tracing is enabled.

describeStateMachineForExecutionResponse_loggingConfiguration :: Lens' DescribeStateMachineForExecutionResponse (Maybe LoggingConfiguration)

Undocumented member.

describeStateMachineForExecutionResponse_httpStatus :: Lens' DescribeStateMachineForExecutionResponse Int

The response's http status code.

describeStateMachineForExecutionResponse_stateMachineArn :: Lens' DescribeStateMachineForExecutionResponse Text

The Amazon Resource Name (ARN) of the state machine associated with the execution.

describeStateMachineForExecutionResponse_name :: Lens' DescribeStateMachineForExecutionResponse Text

The name of the state machine associated with the execution.

describeStateMachineForExecutionResponse_definition :: Lens' DescribeStateMachineForExecutionResponse Text

The Amazon States Language definition of the state machine. See Amazon States Language.

describeStateMachineForExecutionResponse_roleArn :: Lens' DescribeStateMachineForExecutionResponse Text

The Amazon Resource Name (ARN) of the IAM role of the State Machine for the execution.

describeStateMachineForExecutionResponse_updateDate :: Lens' DescribeStateMachineForExecutionResponse UTCTime

The date and time the state machine associated with an execution was updated. For a newly created state machine, this is the creation date.

GetActivityTask

getActivityTask_workerName :: Lens' GetActivityTask (Maybe Text)

You can provide an arbitrary name in order to identify the worker that the task is assigned to. This name is used when it is logged in the execution history.

getActivityTask_activityArn :: Lens' GetActivityTask Text

The Amazon Resource Name (ARN) of the activity to retrieve tasks from (assigned when you create the task using CreateActivity.)

getActivityTaskResponse_input :: Lens' GetActivityTaskResponse (Maybe Text)

The string that contains the JSON input data for the task. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

getActivityTaskResponse_taskToken :: Lens' GetActivityTaskResponse (Maybe Text)

A token that identifies the scheduled task. This token must be copied and included in subsequent calls to SendTaskHeartbeat, SendTaskSuccess or SendTaskFailure in order to report the progress or completion of the task.

getActivityTaskResponse_httpStatus :: Lens' GetActivityTaskResponse Int

The response's http status code.

CreateActivity

createActivity_tags :: Lens' CreateActivity (Maybe [Tag])

The list of tags to add to a resource.

An array of key-value pairs. For more information, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide, and Controlling Access Using IAM Tags.

Tags may only contain Unicode letters, digits, white space, or these symbols: _ . : / = + - @.

createActivity_name :: Lens' CreateActivity Text

The name of the activity to create. This name must be unique for your AWS account and region for 90 days. For more information, see Limits Related to State Machine Executions in the AWS Step Functions Developer Guide.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

createActivityResponse_httpStatus :: Lens' CreateActivityResponse Int

The response's http status code.

createActivityResponse_activityArn :: Lens' CreateActivityResponse Text

The Amazon Resource Name (ARN) that identifies the created activity.

createActivityResponse_creationDate :: Lens' CreateActivityResponse UTCTime

The date the activity is created.

ListActivities

listActivities_nextToken :: Lens' ListActivities (Maybe Text)

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

listActivities_maxResults :: Lens' ListActivities (Maybe Natural)

The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000. A value of 0 uses the default.

This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.

listActivitiesResponse_nextToken :: Lens' ListActivitiesResponse (Maybe Text)

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

listActivitiesResponse_httpStatus :: Lens' ListActivitiesResponse Int

The response's http status code.

listActivitiesResponse_activities :: Lens' ListActivitiesResponse [ActivityListItem]

The list of activities.

SendTaskHeartbeat

sendTaskHeartbeat_taskToken :: Lens' SendTaskHeartbeat Text

The token that represents this task. Task tokens are generated by Step Functions when tasks are assigned to a worker, or in the context object when a workflow enters a task state. See GetActivityTaskOutput$taskToken.

sendTaskHeartbeatResponse_httpStatus :: Lens' SendTaskHeartbeatResponse Int

The response's http status code.

SendTaskFailure

sendTaskFailure_error :: Lens' SendTaskFailure (Maybe Text)

The error code of the failure.

sendTaskFailure_cause :: Lens' SendTaskFailure (Maybe Text)

A more detailed explanation of the cause of the failure.

sendTaskFailure_taskToken :: Lens' SendTaskFailure Text

The token that represents this task. Task tokens are generated by Step Functions when tasks are assigned to a worker, or in the context object when a workflow enters a task state. See GetActivityTaskOutput$taskToken.

sendTaskFailureResponse_httpStatus :: Lens' SendTaskFailureResponse Int

The response's http status code.

DescribeExecution

describeExecution_executionArn :: Lens' DescribeExecution Text

The Amazon Resource Name (ARN) of the execution to describe.

describeExecutionResponse_stopDate :: Lens' DescribeExecutionResponse (Maybe UTCTime)

If the execution has already ended, the date the execution stopped.

describeExecutionResponse_inputDetails :: Lens' DescribeExecutionResponse (Maybe CloudWatchEventsExecutionDataDetails)

Undocumented member.

describeExecutionResponse_input :: Lens' DescribeExecutionResponse (Maybe Text)

The string that contains the JSON input data of the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

describeExecutionResponse_name :: Lens' DescribeExecutionResponse (Maybe Text)

The name of the execution.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

describeExecutionResponse_output :: Lens' DescribeExecutionResponse (Maybe Text)

The JSON output data of the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

This field is set only if the execution succeeds. If the execution fails, this field is null.

describeExecutionResponse_outputDetails :: Lens' DescribeExecutionResponse (Maybe CloudWatchEventsExecutionDataDetails)

Undocumented member.

describeExecutionResponse_traceHeader :: Lens' DescribeExecutionResponse (Maybe Text)

The AWS X-Ray trace header that was passed to the execution.

describeExecutionResponse_httpStatus :: Lens' DescribeExecutionResponse Int

The response's http status code.

describeExecutionResponse_executionArn :: Lens' DescribeExecutionResponse Text

The Amazon Resource Name (ARN) that identifies the execution.

describeExecutionResponse_stateMachineArn :: Lens' DescribeExecutionResponse Text

The Amazon Resource Name (ARN) of the executed stated machine.

describeExecutionResponse_status :: Lens' DescribeExecutionResponse ExecutionStatus

The current status of the execution.

describeExecutionResponse_startDate :: Lens' DescribeExecutionResponse UTCTime

The date the execution is started.

SendTaskSuccess

sendTaskSuccess_taskToken :: Lens' SendTaskSuccess Text

The token that represents this task. Task tokens are generated by Step Functions when tasks are assigned to a worker, or in the context object when a workflow enters a task state. See GetActivityTaskOutput$taskToken.

sendTaskSuccess_output :: Lens' SendTaskSuccess Text

The JSON output of the task. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

sendTaskSuccessResponse_httpStatus :: Lens' SendTaskSuccessResponse Int

The response's http status code.

StartExecution

startExecution_input :: Lens' StartExecution (Maybe Text)

The string that contains the JSON input data for the execution, for example:

"input": "{\"first_name\" : \"test\"}"

If you don't include any JSON input data, you still must include the two braces, for example: "input": "{}"

Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

startExecution_name :: Lens' StartExecution (Maybe Text)

The name of the execution. This name must be unique for your AWS account, region, and state machine for 90 days. For more information, see Limits Related to State Machine Executions in the AWS Step Functions Developer Guide.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

startExecution_traceHeader :: Lens' StartExecution (Maybe Text)

Passes the AWS X-Ray trace header. The trace header can also be passed in the request payload.

startExecution_stateMachineArn :: Lens' StartExecution Text

The Amazon Resource Name (ARN) of the state machine to execute.

startExecutionResponse_httpStatus :: Lens' StartExecutionResponse Int

The response's http status code.

startExecutionResponse_executionArn :: Lens' StartExecutionResponse Text

The Amazon Resource Name (ARN) that identifies the execution.

startExecutionResponse_startDate :: Lens' StartExecutionResponse UTCTime

The date the execution is started.

StartSyncExecution

startSyncExecution_input :: Lens' StartSyncExecution (Maybe Text)

The string that contains the JSON input data for the execution, for example:

"input": "{\"first_name\" : \"test\"}"

If you don't include any JSON input data, you still must include the two braces, for example: "input": "{}"

Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

startSyncExecution_name :: Lens' StartSyncExecution (Maybe Text)

The name of the execution.

startSyncExecution_traceHeader :: Lens' StartSyncExecution (Maybe Text)

Passes the AWS X-Ray trace header. The trace header can also be passed in the request payload.

startSyncExecution_stateMachineArn :: Lens' StartSyncExecution Text

The Amazon Resource Name (ARN) of the state machine to execute.

startSyncExecutionResponse_inputDetails :: Lens' StartSyncExecutionResponse (Maybe CloudWatchEventsExecutionDataDetails)

Undocumented member.

startSyncExecutionResponse_error :: Lens' StartSyncExecutionResponse (Maybe Text)

The error code of the failure.

startSyncExecutionResponse_input :: Lens' StartSyncExecutionResponse (Maybe Text)

The string that contains the JSON input data of the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

startSyncExecutionResponse_cause :: Lens' StartSyncExecutionResponse (Maybe Text)

A more detailed explanation of the cause of the failure.

startSyncExecutionResponse_name :: Lens' StartSyncExecutionResponse (Maybe Text)

The name of the execution.

startSyncExecutionResponse_stateMachineArn :: Lens' StartSyncExecutionResponse (Maybe Text)

The Amazon Resource Name (ARN) that identifies the state machine.

startSyncExecutionResponse_output :: Lens' StartSyncExecutionResponse (Maybe Text)

The JSON output data of the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

This field is set only if the execution succeeds. If the execution fails, this field is null.

startSyncExecutionResponse_outputDetails :: Lens' StartSyncExecutionResponse (Maybe CloudWatchEventsExecutionDataDetails)

Undocumented member.

startSyncExecutionResponse_traceHeader :: Lens' StartSyncExecutionResponse (Maybe Text)

The AWS X-Ray trace header that was passed to the execution.

startSyncExecutionResponse_billingDetails :: Lens' StartSyncExecutionResponse (Maybe BillingDetails)

An object that describes workflow billing details, including billed duration and memory use.

startSyncExecutionResponse_httpStatus :: Lens' StartSyncExecutionResponse Int

The response's http status code.

startSyncExecutionResponse_executionArn :: Lens' StartSyncExecutionResponse Text

The Amazon Resource Name (ARN) that identifies the execution.

startSyncExecutionResponse_startDate :: Lens' StartSyncExecutionResponse UTCTime

The date the execution is started.

startSyncExecutionResponse_stopDate :: Lens' StartSyncExecutionResponse UTCTime

If the execution has already ended, the date the execution stopped.

startSyncExecutionResponse_status :: Lens' StartSyncExecutionResponse SyncExecutionStatus

The current status of the execution.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) for the Step Functions state machine or activity.

tagResource_tags :: Lens' TagResource [Tag]

The list of tags to add to a resource.

Tags may only contain Unicode letters, digits, white space, or these symbols: _ . : / = + - @.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) for the Step Functions state machine or activity.

untagResource_tagKeys :: Lens' UntagResource [Text]

The list of tags to remove from the resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

GetExecutionHistory

getExecutionHistory_reverseOrder :: Lens' GetExecutionHistory (Maybe Bool)

Lists events in descending order of their timeStamp.

getExecutionHistory_includeExecutionData :: Lens' GetExecutionHistory (Maybe Bool)

You can select whether execution data (input or output of a history event) is returned. The default is true.

getExecutionHistory_nextToken :: Lens' GetExecutionHistory (Maybe Text)

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

getExecutionHistory_maxResults :: Lens' GetExecutionHistory (Maybe Natural)

The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000. A value of 0 uses the default.

This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.

getExecutionHistory_executionArn :: Lens' GetExecutionHistory Text

The Amazon Resource Name (ARN) of the execution.

getExecutionHistoryResponse_nextToken :: Lens' GetExecutionHistoryResponse (Maybe Text)

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

getExecutionHistoryResponse_httpStatus :: Lens' GetExecutionHistoryResponse Int

The response's http status code.

getExecutionHistoryResponse_events :: Lens' GetExecutionHistoryResponse [HistoryEvent]

The list of events that occurred in the execution.

CreateStateMachine

createStateMachine_tracingConfiguration :: Lens' CreateStateMachine (Maybe TracingConfiguration)

Selects whether AWS X-Ray tracing is enabled.

createStateMachine_type :: Lens' CreateStateMachine (Maybe StateMachineType)

Determines whether a Standard or Express state machine is created. The default is STANDARD. You cannot update the type of a state machine once it has been created.

createStateMachine_loggingConfiguration :: Lens' CreateStateMachine (Maybe LoggingConfiguration)

Defines what execution history events are logged and where they are logged.

By default, the level is set to OFF. For more information see Log Levels in the AWS Step Functions User Guide.

createStateMachine_tags :: Lens' CreateStateMachine (Maybe [Tag])

Tags to be added when creating a state machine.

An array of key-value pairs. For more information, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide, and Controlling Access Using IAM Tags.

Tags may only contain Unicode letters, digits, white space, or these symbols: _ . : / = + - @.

createStateMachine_name :: Lens' CreateStateMachine Text

The name of the state machine.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

createStateMachine_definition :: Lens' CreateStateMachine Text

The Amazon States Language definition of the state machine. See Amazon States Language.

createStateMachine_roleArn :: Lens' CreateStateMachine Text

The Amazon Resource Name (ARN) of the IAM role to use for this state machine.

createStateMachineResponse_httpStatus :: Lens' CreateStateMachineResponse Int

The response's http status code.

createStateMachineResponse_stateMachineArn :: Lens' CreateStateMachineResponse Text

The Amazon Resource Name (ARN) that identifies the created state machine.

createStateMachineResponse_creationDate :: Lens' CreateStateMachineResponse UTCTime

The date the state machine is created.

Types

ActivityFailedEventDetails

activityFailedEventDetails_error :: Lens' ActivityFailedEventDetails (Maybe Text)

The error code of the failure.

activityFailedEventDetails_cause :: Lens' ActivityFailedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

ActivityListItem

activityListItem_activityArn :: Lens' ActivityListItem Text

The Amazon Resource Name (ARN) that identifies the activity.

activityListItem_name :: Lens' ActivityListItem Text

The name of the activity.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

activityListItem_creationDate :: Lens' ActivityListItem UTCTime

The date the activity is created.

ActivityScheduleFailedEventDetails

activityScheduleFailedEventDetails_error :: Lens' ActivityScheduleFailedEventDetails (Maybe Text)

The error code of the failure.

activityScheduleFailedEventDetails_cause :: Lens' ActivityScheduleFailedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

ActivityScheduledEventDetails

activityScheduledEventDetails_heartbeatInSeconds :: Lens' ActivityScheduledEventDetails (Maybe Integer)

The maximum allowed duration between two heartbeats for the activity task.

activityScheduledEventDetails_inputDetails :: Lens' ActivityScheduledEventDetails (Maybe HistoryEventExecutionDataDetails)

Contains details about the input for an execution history event.

activityScheduledEventDetails_input :: Lens' ActivityScheduledEventDetails (Maybe Text)

The JSON data input to the activity task. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

activityScheduledEventDetails_timeoutInSeconds :: Lens' ActivityScheduledEventDetails (Maybe Integer)

The maximum allowed duration of the activity task.

activityScheduledEventDetails_resource :: Lens' ActivityScheduledEventDetails Text

The Amazon Resource Name (ARN) of the scheduled activity.

ActivityStartedEventDetails

activityStartedEventDetails_workerName :: Lens' ActivityStartedEventDetails (Maybe Text)

The name of the worker that the task is assigned to. These names are provided by the workers when calling GetActivityTask.

ActivitySucceededEventDetails

activitySucceededEventDetails_output :: Lens' ActivitySucceededEventDetails (Maybe Text)

The JSON data output by the activity task. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

activitySucceededEventDetails_outputDetails :: Lens' ActivitySucceededEventDetails (Maybe HistoryEventExecutionDataDetails)

Contains details about the output of an execution history event.

ActivityTimedOutEventDetails

activityTimedOutEventDetails_error :: Lens' ActivityTimedOutEventDetails (Maybe Text)

The error code of the failure.

activityTimedOutEventDetails_cause :: Lens' ActivityTimedOutEventDetails (Maybe Text)

A more detailed explanation of the cause of the timeout.

BillingDetails

billingDetails_billedMemoryUsedInMB :: Lens' BillingDetails (Maybe Natural)

Billed memory consumption of your workflow, in MB.

billingDetails_billedDurationInMilliseconds :: Lens' BillingDetails (Maybe Natural)

Billed duration of your workflow, in milliseconds.

CloudWatchEventsExecutionDataDetails

cloudWatchEventsExecutionDataDetails_included :: Lens' CloudWatchEventsExecutionDataDetails (Maybe Bool)

Indicates whether input or output was included in the response. Always true for API calls.

CloudWatchLogsLogGroup

cloudWatchLogsLogGroup_logGroupArn :: Lens' CloudWatchLogsLogGroup (Maybe Text)

The ARN of the the CloudWatch log group to which you want your logs emitted to. The ARN must end with :*

ExecutionAbortedEventDetails

executionAbortedEventDetails_error :: Lens' ExecutionAbortedEventDetails (Maybe Text)

The error code of the failure.

executionAbortedEventDetails_cause :: Lens' ExecutionAbortedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

ExecutionFailedEventDetails

executionFailedEventDetails_error :: Lens' ExecutionFailedEventDetails (Maybe Text)

The error code of the failure.

executionFailedEventDetails_cause :: Lens' ExecutionFailedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

ExecutionListItem

executionListItem_stopDate :: Lens' ExecutionListItem (Maybe UTCTime)

If the execution already ended, the date the execution stopped.

executionListItem_executionArn :: Lens' ExecutionListItem Text

The Amazon Resource Name (ARN) that identifies the execution.

executionListItem_stateMachineArn :: Lens' ExecutionListItem Text

The Amazon Resource Name (ARN) of the executed state machine.

executionListItem_name :: Lens' ExecutionListItem Text

The name of the execution.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

executionListItem_status :: Lens' ExecutionListItem ExecutionStatus

The current status of the execution.

executionListItem_startDate :: Lens' ExecutionListItem UTCTime

The date the execution started.

ExecutionStartedEventDetails

executionStartedEventDetails_inputDetails :: Lens' ExecutionStartedEventDetails (Maybe HistoryEventExecutionDataDetails)

Contains details about the input for an execution history event.

executionStartedEventDetails_input :: Lens' ExecutionStartedEventDetails (Maybe Text)

The JSON data input to the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

executionStartedEventDetails_roleArn :: Lens' ExecutionStartedEventDetails (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role used for executing AWS Lambda tasks.

ExecutionSucceededEventDetails

executionSucceededEventDetails_output :: Lens' ExecutionSucceededEventDetails (Maybe Text)

The JSON data output by the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

executionSucceededEventDetails_outputDetails :: Lens' ExecutionSucceededEventDetails (Maybe HistoryEventExecutionDataDetails)

Contains details about the output of an execution history event.

ExecutionTimedOutEventDetails

executionTimedOutEventDetails_error :: Lens' ExecutionTimedOutEventDetails (Maybe Text)

The error code of the failure.

executionTimedOutEventDetails_cause :: Lens' ExecutionTimedOutEventDetails (Maybe Text)

A more detailed explanation of the cause of the timeout.

HistoryEvent

historyEvent_mapStateStartedEventDetails :: Lens' HistoryEvent (Maybe MapStateStartedEventDetails)

Contains details about Map state that was started.

historyEvent_taskSubmitFailedEventDetails :: Lens' HistoryEvent (Maybe TaskSubmitFailedEventDetails)

Contains details about a task that where the submit failed.

historyEvent_taskStartedEventDetails :: Lens' HistoryEvent (Maybe TaskStartedEventDetails)

Contains details about a task that was started.

historyEvent_activityStartedEventDetails :: Lens' HistoryEvent (Maybe ActivityStartedEventDetails)

Undocumented member.

historyEvent_taskSubmittedEventDetails :: Lens' HistoryEvent (Maybe TaskSubmittedEventDetails)

Contains details about a submitted task.

historyEvent_lambdaFunctionStartFailedEventDetails :: Lens' HistoryEvent (Maybe LambdaFunctionStartFailedEventDetails)

Contains details about a lambda function that failed to start during an execution.

historyEvent_taskStartFailedEventDetails :: Lens' HistoryEvent (Maybe TaskStartFailedEventDetails)

Contains details about a task that failed to start.

historyEvent_stateExitedEventDetails :: Lens' HistoryEvent (Maybe StateExitedEventDetails)

Undocumented member.

historyEvent_lambdaFunctionSucceededEventDetails :: Lens' HistoryEvent (Maybe LambdaFunctionSucceededEventDetails)

Contains details about a lambda function that terminated successfully during an execution.

historyEvent_taskSucceededEventDetails :: Lens' HistoryEvent (Maybe TaskSucceededEventDetails)

Contains details about a task that succeeded.

historyEvent_activitySucceededEventDetails :: Lens' HistoryEvent (Maybe ActivitySucceededEventDetails)

Undocumented member.

historyEvent_mapIterationAbortedEventDetails :: Lens' HistoryEvent (Maybe MapIterationEventDetails)

Contains details about an iteration of a Map state that was aborted.

historyEvent_mapIterationSucceededEventDetails :: Lens' HistoryEvent (Maybe MapIterationEventDetails)

Contains details about an iteration of a Map state that succeeded.

historyEvent_mapIterationStartedEventDetails :: Lens' HistoryEvent (Maybe MapIterationEventDetails)

Contains details about an iteration of a Map state that was started.

historyEvent_lambdaFunctionTimedOutEventDetails :: Lens' HistoryEvent (Maybe LambdaFunctionTimedOutEventDetails)

Undocumented member.

historyEvent_taskTimedOutEventDetails :: Lens' HistoryEvent (Maybe TaskTimedOutEventDetails)

Contains details about a task that timed out.

historyEvent_activityTimedOutEventDetails :: Lens' HistoryEvent (Maybe ActivityTimedOutEventDetails)

Undocumented member.

historyEvent_executionFailedEventDetails :: Lens' HistoryEvent (Maybe ExecutionFailedEventDetails)

Undocumented member.

historyEvent_executionAbortedEventDetails :: Lens' HistoryEvent (Maybe ExecutionAbortedEventDetails)

Undocumented member.

historyEvent_executionSucceededEventDetails :: Lens' HistoryEvent (Maybe ExecutionSucceededEventDetails)

Undocumented member.

historyEvent_lambdaFunctionScheduledEventDetails :: Lens' HistoryEvent (Maybe LambdaFunctionScheduledEventDetails)

Undocumented member.

historyEvent_taskScheduledEventDetails :: Lens' HistoryEvent (Maybe TaskScheduledEventDetails)

Contains details about a task that was scheduled.

historyEvent_activityScheduledEventDetails :: Lens' HistoryEvent (Maybe ActivityScheduledEventDetails)

Undocumented member.

historyEvent_executionStartedEventDetails :: Lens' HistoryEvent (Maybe ExecutionStartedEventDetails)

Undocumented member.

historyEvent_activityScheduleFailedEventDetails :: Lens' HistoryEvent (Maybe ActivityScheduleFailedEventDetails)

Contains details about an activity schedule event that failed during an execution.

historyEvent_lambdaFunctionScheduleFailedEventDetails :: Lens' HistoryEvent (Maybe LambdaFunctionScheduleFailedEventDetails)

Undocumented member.

historyEvent_stateEnteredEventDetails :: Lens' HistoryEvent (Maybe StateEnteredEventDetails)

Undocumented member.

historyEvent_previousEventId :: Lens' HistoryEvent (Maybe Integer)

The id of the previous event.

historyEvent_activityFailedEventDetails :: Lens' HistoryEvent (Maybe ActivityFailedEventDetails)

Undocumented member.

historyEvent_taskFailedEventDetails :: Lens' HistoryEvent (Maybe TaskFailedEventDetails)

Contains details about the failure of a task.

historyEvent_lambdaFunctionFailedEventDetails :: Lens' HistoryEvent (Maybe LambdaFunctionFailedEventDetails)

Undocumented member.

historyEvent_executionTimedOutEventDetails :: Lens' HistoryEvent (Maybe ExecutionTimedOutEventDetails)

Undocumented member.

historyEvent_mapIterationFailedEventDetails :: Lens' HistoryEvent (Maybe MapIterationEventDetails)

Contains details about an iteration of a Map state that failed.

historyEvent_timestamp :: Lens' HistoryEvent UTCTime

The date and time the event occurred.

historyEvent_type :: Lens' HistoryEvent HistoryEventType

The type of the event.

historyEvent_id :: Lens' HistoryEvent Integer

The id of the event. Events are numbered sequentially, starting at one.

HistoryEventExecutionDataDetails

historyEventExecutionDataDetails_truncated :: Lens' HistoryEventExecutionDataDetails (Maybe Bool)

Indicates whether input or output was truncated in the response. Always false for API calls.

LambdaFunctionFailedEventDetails

lambdaFunctionFailedEventDetails_error :: Lens' LambdaFunctionFailedEventDetails (Maybe Text)

The error code of the failure.

lambdaFunctionFailedEventDetails_cause :: Lens' LambdaFunctionFailedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

LambdaFunctionScheduleFailedEventDetails

lambdaFunctionScheduleFailedEventDetails_error :: Lens' LambdaFunctionScheduleFailedEventDetails (Maybe Text)

The error code of the failure.

lambdaFunctionScheduleFailedEventDetails_cause :: Lens' LambdaFunctionScheduleFailedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

LambdaFunctionScheduledEventDetails

lambdaFunctionScheduledEventDetails_inputDetails :: Lens' LambdaFunctionScheduledEventDetails (Maybe HistoryEventExecutionDataDetails)

Contains details about input for an execution history event.

lambdaFunctionScheduledEventDetails_input :: Lens' LambdaFunctionScheduledEventDetails (Maybe Text)

The JSON data input to the lambda function. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

lambdaFunctionScheduledEventDetails_timeoutInSeconds :: Lens' LambdaFunctionScheduledEventDetails (Maybe Integer)

The maximum allowed duration of the lambda function.

lambdaFunctionScheduledEventDetails_resource :: Lens' LambdaFunctionScheduledEventDetails Text

The Amazon Resource Name (ARN) of the scheduled lambda function.

LambdaFunctionStartFailedEventDetails

lambdaFunctionStartFailedEventDetails_error :: Lens' LambdaFunctionStartFailedEventDetails (Maybe Text)

The error code of the failure.

lambdaFunctionStartFailedEventDetails_cause :: Lens' LambdaFunctionStartFailedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

LambdaFunctionSucceededEventDetails

lambdaFunctionSucceededEventDetails_output :: Lens' LambdaFunctionSucceededEventDetails (Maybe Text)

The JSON data output by the lambda function. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

lambdaFunctionSucceededEventDetails_outputDetails :: Lens' LambdaFunctionSucceededEventDetails (Maybe HistoryEventExecutionDataDetails)

Contains details about the output of an execution history event.

LambdaFunctionTimedOutEventDetails

lambdaFunctionTimedOutEventDetails_error :: Lens' LambdaFunctionTimedOutEventDetails (Maybe Text)

The error code of the failure.

lambdaFunctionTimedOutEventDetails_cause :: Lens' LambdaFunctionTimedOutEventDetails (Maybe Text)

A more detailed explanation of the cause of the timeout.

LogDestination

logDestination_cloudWatchLogsLogGroup :: Lens' LogDestination (Maybe CloudWatchLogsLogGroup)

An object describing a CloudWatch log group. For more information, see AWS::Logs::LogGroup in the AWS CloudFormation User Guide.

LoggingConfiguration

loggingConfiguration_includeExecutionData :: Lens' LoggingConfiguration (Maybe Bool)

Determines whether execution data is included in your log. When set to false, data is excluded.

loggingConfiguration_destinations :: Lens' LoggingConfiguration (Maybe [LogDestination])

An array of objects that describes where your execution history events will be logged. Limited to size 1. Required, if your log level is not set to OFF.

loggingConfiguration_level :: Lens' LoggingConfiguration (Maybe LogLevel)

Defines which category of execution history events are logged.

MapIterationEventDetails

mapIterationEventDetails_name :: Lens' MapIterationEventDetails (Maybe Text)

The name of the iteration’s parent Map state.

mapIterationEventDetails_index :: Lens' MapIterationEventDetails (Maybe Natural)

The index of the array belonging to the Map state iteration.

MapStateStartedEventDetails

mapStateStartedEventDetails_length :: Lens' MapStateStartedEventDetails (Maybe Natural)

The size of the array for Map state iterations.

StateEnteredEventDetails

stateEnteredEventDetails_inputDetails :: Lens' StateEnteredEventDetails (Maybe HistoryEventExecutionDataDetails)

Contains details about the input for an execution history event.

stateEnteredEventDetails_input :: Lens' StateEnteredEventDetails (Maybe Text)

The string that contains the JSON input data for the state. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

stateEnteredEventDetails_name :: Lens' StateEnteredEventDetails Text

The name of the state.

StateExitedEventDetails

stateExitedEventDetails_output :: Lens' StateExitedEventDetails (Maybe Text)

The JSON output data of the state. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

stateExitedEventDetails_outputDetails :: Lens' StateExitedEventDetails (Maybe HistoryEventExecutionDataDetails)

Contains details about the output of an execution history event.

stateExitedEventDetails_name :: Lens' StateExitedEventDetails Text

The name of the state.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

StateMachineListItem

stateMachineListItem_stateMachineArn :: Lens' StateMachineListItem Text

The Amazon Resource Name (ARN) that identifies the state machine.

stateMachineListItem_name :: Lens' StateMachineListItem Text

The name of the state machine.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

stateMachineListItem_type :: Lens' StateMachineListItem StateMachineType

stateMachineListItem_creationDate :: Lens' StateMachineListItem UTCTime

The date the state machine is created.

Tag

tag_value :: Lens' Tag (Maybe Text)

The value of a tag.

tag_key :: Lens' Tag (Maybe Text)

The key of a tag.

TaskFailedEventDetails

taskFailedEventDetails_error :: Lens' TaskFailedEventDetails (Maybe Text)

The error code of the failure.

taskFailedEventDetails_cause :: Lens' TaskFailedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

taskFailedEventDetails_resourceType :: Lens' TaskFailedEventDetails Text

The action of the resource called by a task state.

taskFailedEventDetails_resource :: Lens' TaskFailedEventDetails Text

The service name of the resource in a task state.

TaskScheduledEventDetails

taskScheduledEventDetails_heartbeatInSeconds :: Lens' TaskScheduledEventDetails (Maybe Integer)

The maximum allowed duration between two heartbeats for the task.

taskScheduledEventDetails_timeoutInSeconds :: Lens' TaskScheduledEventDetails (Maybe Integer)

The maximum allowed duration of the task.

taskScheduledEventDetails_resourceType :: Lens' TaskScheduledEventDetails Text

The action of the resource called by a task state.

taskScheduledEventDetails_resource :: Lens' TaskScheduledEventDetails Text

The service name of the resource in a task state.

taskScheduledEventDetails_region :: Lens' TaskScheduledEventDetails Text

The region of the scheduled task

taskScheduledEventDetails_parameters :: Lens' TaskScheduledEventDetails Text

The JSON data passed to the resource referenced in a task state. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

TaskStartFailedEventDetails

taskStartFailedEventDetails_error :: Lens' TaskStartFailedEventDetails (Maybe Text)

The error code of the failure.

taskStartFailedEventDetails_cause :: Lens' TaskStartFailedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

taskStartFailedEventDetails_resourceType :: Lens' TaskStartFailedEventDetails Text

The action of the resource called by a task state.

taskStartFailedEventDetails_resource :: Lens' TaskStartFailedEventDetails Text

The service name of the resource in a task state.

TaskStartedEventDetails

taskStartedEventDetails_resourceType :: Lens' TaskStartedEventDetails Text

The action of the resource called by a task state.

taskStartedEventDetails_resource :: Lens' TaskStartedEventDetails Text

The service name of the resource in a task state.

TaskSubmitFailedEventDetails

taskSubmitFailedEventDetails_error :: Lens' TaskSubmitFailedEventDetails (Maybe Text)

The error code of the failure.

taskSubmitFailedEventDetails_cause :: Lens' TaskSubmitFailedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

taskSubmitFailedEventDetails_resourceType :: Lens' TaskSubmitFailedEventDetails Text

The action of the resource called by a task state.

taskSubmitFailedEventDetails_resource :: Lens' TaskSubmitFailedEventDetails Text

The service name of the resource in a task state.

TaskSubmittedEventDetails

taskSubmittedEventDetails_output :: Lens' TaskSubmittedEventDetails (Maybe Text)

The response from a resource when a task has started. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

taskSubmittedEventDetails_outputDetails :: Lens' TaskSubmittedEventDetails (Maybe HistoryEventExecutionDataDetails)

Contains details about the output of an execution history event.

taskSubmittedEventDetails_resourceType :: Lens' TaskSubmittedEventDetails Text

The action of the resource called by a task state.

taskSubmittedEventDetails_resource :: Lens' TaskSubmittedEventDetails Text

The service name of the resource in a task state.

TaskSucceededEventDetails

taskSucceededEventDetails_output :: Lens' TaskSucceededEventDetails (Maybe Text)

The full JSON response from a resource when a task has succeeded. This response becomes the output of the related task. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

taskSucceededEventDetails_outputDetails :: Lens' TaskSucceededEventDetails (Maybe HistoryEventExecutionDataDetails)

Contains details about the output of an execution history event.

taskSucceededEventDetails_resourceType :: Lens' TaskSucceededEventDetails Text

The action of the resource called by a task state.

taskSucceededEventDetails_resource :: Lens' TaskSucceededEventDetails Text

The service name of the resource in a task state.

TaskTimedOutEventDetails

taskTimedOutEventDetails_error :: Lens' TaskTimedOutEventDetails (Maybe Text)

The error code of the failure.

taskTimedOutEventDetails_cause :: Lens' TaskTimedOutEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

taskTimedOutEventDetails_resourceType :: Lens' TaskTimedOutEventDetails Text

The action of the resource called by a task state.

taskTimedOutEventDetails_resource :: Lens' TaskTimedOutEventDetails Text

The service name of the resource in a task state.

TracingConfiguration

tracingConfiguration_enabled :: Lens' TracingConfiguration (Maybe Bool)

When set to true, AWS X-Ray tracing is enabled.