Network.AWS.StepFunctions.ListExecutions

Description

Lists the executions of a state machine that meet the filtering criteria. Results are sorted by time, with the most recent execution first.

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

This API action is not supported by EXPRESS state machines.

This operation returns paginated results.

Creating a Request

data ListExecutions

See: newListExecutions smart constructor.

Constructors

ListExecutions'

Fields

nextToken :: Maybe Text

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

maxResults :: Maybe Natural

The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000. A value of 0 uses the default.

This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.

statusFilter :: Maybe ExecutionStatus

If specified, only list the executions whose current execution status matches the given filter.

stateMachineArn :: Text

The Amazon Resource Name (ARN) of the state machine whose executions is listed.

newListExecutions

Arguments

:: Text
$sel:stateMachineArn:ListExecutions'

-> ListExecutions

Create a value of ListExecutions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListExecutions', listExecutions_nextToken - If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

$sel:maxResults:ListExecutions', listExecutions_maxResults - The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000. A value of 0 uses the default.

This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.

$sel:statusFilter:ListExecutions', listExecutions_statusFilter - If specified, only list the executions whose current execution status matches the given filter.

$sel:stateMachineArn:ListExecutions', listExecutions_stateMachineArn - The Amazon Resource Name (ARN) of the state machine whose executions is listed.

Request Lenses

listExecutions_nextToken :: Lens' ListExecutions (Maybe Text)

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

listExecutions_maxResults :: Lens' ListExecutions (Maybe Natural)

The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000. A value of 0 uses the default.

This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.

listExecutions_statusFilter :: Lens' ListExecutions (Maybe ExecutionStatus)

If specified, only list the executions whose current execution status matches the given filter.

listExecutions_stateMachineArn :: Lens' ListExecutions Text

The Amazon Resource Name (ARN) of the state machine whose executions is listed.

Destructuring the Response

data ListExecutionsResponse

See: newListExecutionsResponse smart constructor.

Constructors

ListExecutionsResponse'

Fields

nextToken :: Maybe Text

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

httpStatus :: Int

The response's http status code.

executions :: [ExecutionListItem]

The list of matching executions.

newListExecutionsResponse

Arguments

:: Int
$sel:httpStatus:ListExecutionsResponse'

-> ListExecutionsResponse

Create a value of ListExecutionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListExecutions', listExecutionsResponse_nextToken - If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

$sel:httpStatus:ListExecutionsResponse', listExecutionsResponse_httpStatus - The response's http status code.

$sel:executions:ListExecutionsResponse', listExecutionsResponse_executions - The list of matching executions.

Response Lenses

listExecutionsResponse_nextToken :: Lens' ListExecutionsResponse (Maybe Text)

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

listExecutionsResponse_httpStatus :: Lens' ListExecutionsResponse Int

The response's http status code.

listExecutionsResponse_executions :: Lens' ListExecutionsResponse [ExecutionListItem]

The list of matching executions.