Network.AWS.StepFunctions.Types.LambdaFunctionFailedEventDetails

Description

Documentation

data LambdaFunctionFailedEventDetails

Contains details about a lambda function that failed during an execution.

See: newLambdaFunctionFailedEventDetails smart constructor.

Constructors

LambdaFunctionFailedEventDetails'

Fields

cause :: Maybe (Sensitive Text)

A more detailed explanation of the cause of the failure.

error :: Maybe (Sensitive Text)

The error code of the failure.

newLambdaFunctionFailedEventDetails :: LambdaFunctionFailedEventDetails

Create a value of LambdaFunctionFailedEventDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cause:LambdaFunctionFailedEventDetails', lambdaFunctionFailedEventDetails_cause - A more detailed explanation of the cause of the failure.

$sel:error:LambdaFunctionFailedEventDetails', lambdaFunctionFailedEventDetails_error - The error code of the failure.

lambdaFunctionFailedEventDetails_cause :: Lens' LambdaFunctionFailedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

lambdaFunctionFailedEventDetails_error :: Lens' LambdaFunctionFailedEventDetails (Maybe Text)

The error code of the failure.