Network.AWS.StepFunctions.Types.TaskStartFailedEventDetails

Description

Documentation

data TaskStartFailedEventDetails

Contains details about a task that failed to start during an execution.

See: newTaskStartFailedEventDetails smart constructor.

Constructors

TaskStartFailedEventDetails'

Fields

cause :: Maybe (Sensitive Text)

A more detailed explanation of the cause of the failure.

error :: Maybe (Sensitive Text)

The error code of the failure.

resourceType :: Text

The action of the resource called by a task state.

resource :: Text

The service name of the resource in a task state.

newTaskStartFailedEventDetails

Arguments

:: Text
$sel:resourceType:TaskStartFailedEventDetails'

-> Text
$sel:resource:TaskStartFailedEventDetails'

-> TaskStartFailedEventDetails

Create a value of TaskStartFailedEventDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cause:TaskStartFailedEventDetails', taskStartFailedEventDetails_cause - A more detailed explanation of the cause of the failure.

$sel:error:TaskStartFailedEventDetails', taskStartFailedEventDetails_error - The error code of the failure.

$sel:resourceType:TaskStartFailedEventDetails', taskStartFailedEventDetails_resourceType - The action of the resource called by a task state.

$sel:resource:TaskStartFailedEventDetails', taskStartFailedEventDetails_resource - The service name of the resource in a task state.

taskStartFailedEventDetails_cause :: Lens' TaskStartFailedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

taskStartFailedEventDetails_error :: Lens' TaskStartFailedEventDetails (Maybe Text)

The error code of the failure.

taskStartFailedEventDetails_resourceType :: Lens' TaskStartFailedEventDetails Text

The action of the resource called by a task state.

taskStartFailedEventDetails_resource :: Lens' TaskStartFailedEventDetails Text

The service name of the resource in a task state.