Network.AWS.StepFunctions.Types.TaskSubmittedEventDetails

Description

Documentation

data TaskSubmittedEventDetails

Contains details about a task submitted to a resource .

See: newTaskSubmittedEventDetails smart constructor.

Constructors

TaskSubmittedEventDetails'

Fields

output :: Maybe (Sensitive Text)

The response from a resource when a task has started. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

outputDetails :: Maybe HistoryEventExecutionDataDetails

Contains details about the output of an execution history event.

resourceType :: Text

The action of the resource called by a task state.

resource :: Text

The service name of the resource in a task state.

newTaskSubmittedEventDetails

Arguments

:: Text
$sel:resourceType:TaskSubmittedEventDetails'

-> Text
$sel:resource:TaskSubmittedEventDetails'

-> TaskSubmittedEventDetails

Create a value of TaskSubmittedEventDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:output:TaskSubmittedEventDetails', taskSubmittedEventDetails_output - The response from a resource when a task has started. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

$sel:outputDetails:TaskSubmittedEventDetails', taskSubmittedEventDetails_outputDetails - Contains details about the output of an execution history event.

$sel:resourceType:TaskSubmittedEventDetails', taskSubmittedEventDetails_resourceType - The action of the resource called by a task state.

$sel:resource:TaskSubmittedEventDetails', taskSubmittedEventDetails_resource - The service name of the resource in a task state.

taskSubmittedEventDetails_output :: Lens' TaskSubmittedEventDetails (Maybe Text)

The response from a resource when a task has started. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

taskSubmittedEventDetails_outputDetails :: Lens' TaskSubmittedEventDetails (Maybe HistoryEventExecutionDataDetails)

Contains details about the output of an execution history event.

taskSubmittedEventDetails_resourceType :: Lens' TaskSubmittedEventDetails Text

The action of the resource called by a task state.

taskSubmittedEventDetails_resource :: Lens' TaskSubmittedEventDetails Text

The service name of the resource in a task state.