Network.AWS.StorageGateway.AddUploadBuffer

Description

Configures one or more gateway local disks as upload buffer for a specified gateway. This operation is supported for the stored volume, cached volume, and tape gateway types.

In the request, you specify the gateway Amazon Resource Name (ARN) to which you want to add upload buffer, and one or more disk IDs that you want to configure as upload buffer.

Creating a Request

data AddUploadBuffer

See: newAddUploadBuffer smart constructor.

Constructors

AddUploadBuffer'

Fields

gatewayARN :: Text

diskIds :: [Text]

An array of strings that identify disks that are to be configured as working storage. Each string has a minimum length of 1 and maximum length of 300. You can get the disk IDs from the ListLocalDisks API.

newAddUploadBuffer

Arguments

:: Text
$sel:gatewayARN:AddUploadBuffer'

-> AddUploadBuffer

Create a value of AddUploadBuffer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:AddUploadBuffer', addUploadBuffer_gatewayARN - Undocumented member.

$sel:diskIds:AddUploadBuffer', addUploadBuffer_diskIds - An array of strings that identify disks that are to be configured as working storage. Each string has a minimum length of 1 and maximum length of 300. You can get the disk IDs from the ListLocalDisks API.

Request Lenses

addUploadBuffer_gatewayARN :: Lens' AddUploadBuffer Text

Undocumented member.

addUploadBuffer_diskIds :: Lens' AddUploadBuffer [Text]

An array of strings that identify disks that are to be configured as working storage. Each string has a minimum length of 1 and maximum length of 300. You can get the disk IDs from the ListLocalDisks API.

Destructuring the Response

data AddUploadBufferResponse

See: newAddUploadBufferResponse smart constructor.

Constructors

AddUploadBufferResponse'

Fields

gatewayARN :: Maybe Text

httpStatus :: Int

The response's http status code.

newAddUploadBufferResponse

Arguments

:: Int
$sel:httpStatus:AddUploadBufferResponse'

-> AddUploadBufferResponse

Create a value of AddUploadBufferResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:AddUploadBuffer', addUploadBufferResponse_gatewayARN - Undocumented member.

$sel:httpStatus:AddUploadBufferResponse', addUploadBufferResponse_httpStatus - The response's http status code.

Response Lenses

addUploadBufferResponse_gatewayARN :: Lens' AddUploadBufferResponse (Maybe Text)

Undocumented member.

addUploadBufferResponse_httpStatus :: Lens' AddUploadBufferResponse Int

The response's http status code.