Network.AWS.StorageGateway.AssignTapePool

Description

Assigns a tape to a tape pool for archiving. The tape assigned to a pool is archived in the S3 storage class that is associated with the pool. When you use your backup application to eject the tape, the tape is archived directly into the S3 storage class (S3 Glacier or S3 Glacier Deep Archive) that corresponds to the pool.

Valid Values: GLACIER | DEEP_ARCHIVE

Creating a Request

data AssignTapePool

See: newAssignTapePool smart constructor.

Constructors

AssignTapePool'

Fields

bypassGovernanceRetention :: Maybe Bool

Set permissions to bypass governance retention. If the lock type of the archived tape is Governance, the tape's archived age is not older than RetentionLockInDays, and the user does not already have BypassGovernanceRetention, setting this to TRUE enables the user to bypass the retention lock. This parameter is set to true by default for calls from the console.

Valid values: TRUE | FALSE

tapeARN :: Text

The unique Amazon Resource Name (ARN) of the virtual tape that you want to add to the tape pool.

poolId :: Text

The ID of the pool that you want to add your tape to for archiving. The tape in this pool is archived in the S3 storage class that is associated with the pool. When you use your backup application to eject the tape, the tape is archived directly into the storage class (S3 Glacier or S3 Glacier Deep Archive) that corresponds to the pool.

Valid Values: GLACIER | DEEP_ARCHIVE

newAssignTapePool

Arguments

:: Text
$sel:tapeARN:AssignTapePool'

-> Text
$sel:poolId:AssignTapePool'

-> AssignTapePool

Create a value of AssignTapePool with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bypassGovernanceRetention:AssignTapePool', assignTapePool_bypassGovernanceRetention - Set permissions to bypass governance retention. If the lock type of the archived tape is Governance, the tape's archived age is not older than RetentionLockInDays, and the user does not already have BypassGovernanceRetention, setting this to TRUE enables the user to bypass the retention lock. This parameter is set to true by default for calls from the console.

Valid values: TRUE | FALSE

$sel:tapeARN:AssignTapePool', assignTapePool_tapeARN - The unique Amazon Resource Name (ARN) of the virtual tape that you want to add to the tape pool.

$sel:poolId:AssignTapePool', assignTapePool_poolId - The ID of the pool that you want to add your tape to for archiving. The tape in this pool is archived in the S3 storage class that is associated with the pool. When you use your backup application to eject the tape, the tape is archived directly into the storage class (S3 Glacier or S3 Glacier Deep Archive) that corresponds to the pool.

Valid Values: GLACIER | DEEP_ARCHIVE

Request Lenses

assignTapePool_bypassGovernanceRetention :: Lens' AssignTapePool (Maybe Bool)

Set permissions to bypass governance retention. If the lock type of the archived tape is Governance, the tape's archived age is not older than RetentionLockInDays, and the user does not already have BypassGovernanceRetention, setting this to TRUE enables the user to bypass the retention lock. This parameter is set to true by default for calls from the console.

Valid values: TRUE | FALSE

assignTapePool_tapeARN :: Lens' AssignTapePool Text

The unique Amazon Resource Name (ARN) of the virtual tape that you want to add to the tape pool.

assignTapePool_poolId :: Lens' AssignTapePool Text

The ID of the pool that you want to add your tape to for archiving. The tape in this pool is archived in the S3 storage class that is associated with the pool. When you use your backup application to eject the tape, the tape is archived directly into the storage class (S3 Glacier or S3 Glacier Deep Archive) that corresponds to the pool.

Valid Values: GLACIER | DEEP_ARCHIVE

Destructuring the Response

data AssignTapePoolResponse

See: newAssignTapePoolResponse smart constructor.

Constructors

AssignTapePoolResponse'

Fields

tapeARN :: Maybe Text

The unique Amazon Resource Names (ARN) of the virtual tape that was added to the tape pool.

httpStatus :: Int

The response's http status code.

newAssignTapePoolResponse

Arguments

:: Int
$sel:httpStatus:AssignTapePoolResponse'

-> AssignTapePoolResponse

Create a value of AssignTapePoolResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tapeARN:AssignTapePool', assignTapePoolResponse_tapeARN - The unique Amazon Resource Names (ARN) of the virtual tape that was added to the tape pool.

$sel:httpStatus:AssignTapePoolResponse', assignTapePoolResponse_httpStatus - The response's http status code.

Response Lenses

assignTapePoolResponse_tapeARN :: Lens' AssignTapePoolResponse (Maybe Text)

The unique Amazon Resource Names (ARN) of the virtual tape that was added to the tape pool.

assignTapePoolResponse_httpStatus :: Lens' AssignTapePoolResponse Int

The response's http status code.