Network.AWS.StorageGateway.CancelRetrieval

Description

Cancels retrieval of a virtual tape from the virtual tape shelf (VTS) to a gateway after the retrieval process is initiated. The virtual tape is returned to the VTS. This operation is only supported in the tape gateway type.

Creating a Request

data CancelRetrieval

CancelRetrievalInput

See: newCancelRetrieval smart constructor.

Constructors

CancelRetrieval'

Fields

gatewayARN :: Text

tapeARN :: Text

The Amazon Resource Name (ARN) of the virtual tape you want to cancel retrieval for.

newCancelRetrieval

Arguments

:: Text
$sel:gatewayARN:CancelRetrieval'

-> Text
$sel:tapeARN:CancelRetrieval'

-> CancelRetrieval

Create a value of CancelRetrieval with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:CancelRetrieval', cancelRetrieval_gatewayARN - Undocumented member.

$sel:tapeARN:CancelRetrieval', cancelRetrieval_tapeARN - The Amazon Resource Name (ARN) of the virtual tape you want to cancel retrieval for.

Request Lenses

cancelRetrieval_gatewayARN :: Lens' CancelRetrieval Text

Undocumented member.

cancelRetrieval_tapeARN :: Lens' CancelRetrieval Text

The Amazon Resource Name (ARN) of the virtual tape you want to cancel retrieval for.

Destructuring the Response

data CancelRetrievalResponse

CancelRetrievalOutput

See: newCancelRetrievalResponse smart constructor.

Constructors

CancelRetrievalResponse'

Fields

tapeARN :: Maybe Text

The Amazon Resource Name (ARN) of the virtual tape for which retrieval was canceled.

httpStatus :: Int

The response's http status code.

newCancelRetrievalResponse

Arguments

:: Int
$sel:httpStatus:CancelRetrievalResponse'

-> CancelRetrievalResponse

Create a value of CancelRetrievalResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tapeARN:CancelRetrieval', cancelRetrievalResponse_tapeARN - The Amazon Resource Name (ARN) of the virtual tape for which retrieval was canceled.

$sel:httpStatus:CancelRetrievalResponse', cancelRetrievalResponse_httpStatus - The response's http status code.

Response Lenses

cancelRetrievalResponse_tapeARN :: Lens' CancelRetrievalResponse (Maybe Text)

The Amazon Resource Name (ARN) of the virtual tape for which retrieval was canceled.

cancelRetrievalResponse_httpStatus :: Lens' CancelRetrievalResponse Int

The response's http status code.