Network.AWS.StorageGateway.CreateSnapshot

Description

Initiates a snapshot of a volume.

Storage Gateway provides the ability to back up point-in-time snapshots of your data to Amazon Simple Storage (Amazon S3) for durable off-site recovery, and also import the data to an Amazon Elastic Block Store (EBS) volume in Amazon Elastic Compute Cloud (EC2). You can take snapshots of your gateway volume on a scheduled or ad hoc basis. This API enables you to take an ad hoc snapshot. For more information, see Editing a snapshot schedule.

In the CreateSnapshot request, you identify the volume by providing its Amazon Resource Name (ARN). You must also provide description for the snapshot. When Storage Gateway takes the snapshot of specified volume, the snapshot and description appears in the Storage Gateway console. In response, Storage Gateway returns you a snapshot ID. You can use this snapshot ID to check the snapshot progress or later use it when you want to create a volume from a snapshot. This operation is only supported in stored and cached volume gateway type.

To list or delete a snapshot, you must use the Amazon EC2 API. For more information, see DescribeSnapshots or DeleteSnapshot in the Amazon Elastic Compute Cloud API Reference.

Volume and snapshot IDs are changing to a longer length ID format. For more information, see the important note on the Welcome page.

Creating a Request

data CreateSnapshot

A JSON object containing one or more of the following fields:

CreateSnapshotInput$SnapshotDescription

CreateSnapshotInput$VolumeARN

See: newCreateSnapshot smart constructor.

Constructors

CreateSnapshot'

Fields

tags :: Maybe [Tag]

A list of up to 50 tags that can be assigned to a snapshot. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

volumeARN :: Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

snapshotDescription :: Text

Textual description of the snapshot that appears in the Amazon EC2 console, Elastic Block Store snapshots panel in the Description field, and in the Storage Gateway snapshot Details pane, Description field.

newCreateSnapshot

Arguments

:: Text
$sel:volumeARN:CreateSnapshot'

-> Text
$sel:snapshotDescription:CreateSnapshot'

-> CreateSnapshot

Create a value of CreateSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateSnapshot', createSnapshot_tags - A list of up to 50 tags that can be assigned to a snapshot. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

$sel:volumeARN:CreateSnapshot', createSnapshot_volumeARN - The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

$sel:snapshotDescription:CreateSnapshot', createSnapshot_snapshotDescription - Textual description of the snapshot that appears in the Amazon EC2 console, Elastic Block Store snapshots panel in the Description field, and in the Storage Gateway snapshot Details pane, Description field.

Request Lenses

createSnapshot_tags :: Lens' CreateSnapshot (Maybe [Tag])

A list of up to 50 tags that can be assigned to a snapshot. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

createSnapshot_volumeARN :: Lens' CreateSnapshot Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

createSnapshot_snapshotDescription :: Lens' CreateSnapshot Text

Textual description of the snapshot that appears in the Amazon EC2 console, Elastic Block Store snapshots panel in the Description field, and in the Storage Gateway snapshot Details pane, Description field.

Destructuring the Response

data CreateSnapshotResponse

A JSON object containing the following fields:

See: newCreateSnapshotResponse smart constructor.

Constructors

CreateSnapshotResponse'

Fields

volumeARN :: Maybe Text

The Amazon Resource Name (ARN) of the volume of which the snapshot was taken.

snapshotId :: Maybe Text

The snapshot ID that is used to refer to the snapshot in future operations such as describing snapshots (Amazon Elastic Compute Cloud API DescribeSnapshots) or creating a volume from a snapshot (CreateStorediSCSIVolume).

httpStatus :: Int

The response's http status code.

newCreateSnapshotResponse

Arguments

:: Int
$sel:httpStatus:CreateSnapshotResponse'

-> CreateSnapshotResponse

Create a value of CreateSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:volumeARN:CreateSnapshot', createSnapshotResponse_volumeARN - The Amazon Resource Name (ARN) of the volume of which the snapshot was taken.

$sel:snapshotId:CreateSnapshotResponse', createSnapshotResponse_snapshotId - The snapshot ID that is used to refer to the snapshot in future operations such as describing snapshots (Amazon Elastic Compute Cloud API DescribeSnapshots) or creating a volume from a snapshot (CreateStorediSCSIVolume).

$sel:httpStatus:CreateSnapshotResponse', createSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

createSnapshotResponse_volumeARN :: Lens' CreateSnapshotResponse (Maybe Text)

The Amazon Resource Name (ARN) of the volume of which the snapshot was taken.

createSnapshotResponse_snapshotId :: Lens' CreateSnapshotResponse (Maybe Text)

The snapshot ID that is used to refer to the snapshot in future operations such as describing snapshots (Amazon Elastic Compute Cloud API DescribeSnapshots) or creating a volume from a snapshot (CreateStorediSCSIVolume).

createSnapshotResponse_httpStatus :: Lens' CreateSnapshotResponse Int

The response's http status code.