Network.AWS.StorageGateway.CreateSnapshotFromVolumeRecoveryPoint

Description

Initiates a snapshot of a gateway from a volume recovery point. This operation is only supported in the cached volume gateway type.

A volume recovery point is a point in time at which all data of the volume is consistent and from which you can create a snapshot. To get a list of volume recovery point for cached volume gateway, use ListVolumeRecoveryPoints.

In the CreateSnapshotFromVolumeRecoveryPoint request, you identify the volume by providing its Amazon Resource Name (ARN). You must also provide a description for the snapshot. When the gateway takes a snapshot of the specified volume, the snapshot and its description appear in the Storage Gateway console. In response, the gateway returns you a snapshot ID. You can use this snapshot ID to check the snapshot progress or later use it when you want to create a volume from a snapshot.

To list or delete a snapshot, you must use the Amazon EC2 API. For more information, see DescribeSnapshots or DeleteSnapshot in the Amazon Elastic Compute Cloud API Reference.

Creating a Request

data CreateSnapshotFromVolumeRecoveryPoint

See: newCreateSnapshotFromVolumeRecoveryPoint smart constructor.

Constructors

CreateSnapshotFromVolumeRecoveryPoint'

Fields

tags :: Maybe [Tag]

A list of up to 50 tags that can be assigned to a snapshot. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

volumeARN :: Text

The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

snapshotDescription :: Text

Textual description of the snapshot that appears in the Amazon EC2 console, Elastic Block Store snapshots panel in the Description field, and in the Storage Gateway snapshot Details pane, Description field.

newCreateSnapshotFromVolumeRecoveryPoint

Arguments

:: Text
$sel:volumeARN:CreateSnapshotFromVolumeRecoveryPoint'

-> Text
$sel:snapshotDescription:CreateSnapshotFromVolumeRecoveryPoint'

-> CreateSnapshotFromVolumeRecoveryPoint

Create a value of CreateSnapshotFromVolumeRecoveryPoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateSnapshotFromVolumeRecoveryPoint', createSnapshotFromVolumeRecoveryPoint_tags - A list of up to 50 tags that can be assigned to a snapshot. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

$sel:volumeARN:CreateSnapshotFromVolumeRecoveryPoint', createSnapshotFromVolumeRecoveryPoint_volumeARN - The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

$sel:snapshotDescription:CreateSnapshotFromVolumeRecoveryPoint', createSnapshotFromVolumeRecoveryPoint_snapshotDescription - Textual description of the snapshot that appears in the Amazon EC2 console, Elastic Block Store snapshots panel in the Description field, and in the Storage Gateway snapshot Details pane, Description field.

Request Lenses

createSnapshotFromVolumeRecoveryPoint_tags :: Lens' CreateSnapshotFromVolumeRecoveryPoint (Maybe [Tag])

A list of up to 50 tags that can be assigned to a snapshot. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

createSnapshotFromVolumeRecoveryPoint_volumeARN :: Lens' CreateSnapshotFromVolumeRecoveryPoint Text

The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

createSnapshotFromVolumeRecoveryPoint_snapshotDescription :: Lens' CreateSnapshotFromVolumeRecoveryPoint Text

Textual description of the snapshot that appears in the Amazon EC2 console, Elastic Block Store snapshots panel in the Description field, and in the Storage Gateway snapshot Details pane, Description field.

Destructuring the Response

data CreateSnapshotFromVolumeRecoveryPointResponse

See: newCreateSnapshotFromVolumeRecoveryPointResponse smart constructor.

Constructors

CreateSnapshotFromVolumeRecoveryPointResponse'

Fields

volumeRecoveryPointTime :: Maybe Text

The time the volume was created from the recovery point.

volumeARN :: Maybe Text

The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

snapshotId :: Maybe Text

The ID of the snapshot.

httpStatus :: Int

The response's http status code.

newCreateSnapshotFromVolumeRecoveryPointResponse

Arguments

:: Int
$sel:httpStatus:CreateSnapshotFromVolumeRecoveryPointResponse'

-> CreateSnapshotFromVolumeRecoveryPointResponse

Create a value of CreateSnapshotFromVolumeRecoveryPointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:volumeRecoveryPointTime:CreateSnapshotFromVolumeRecoveryPointResponse', createSnapshotFromVolumeRecoveryPointResponse_volumeRecoveryPointTime - The time the volume was created from the recovery point.

$sel:volumeARN:CreateSnapshotFromVolumeRecoveryPoint', createSnapshotFromVolumeRecoveryPointResponse_volumeARN - The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

$sel:snapshotId:CreateSnapshotFromVolumeRecoveryPointResponse', createSnapshotFromVolumeRecoveryPointResponse_snapshotId - The ID of the snapshot.

$sel:httpStatus:CreateSnapshotFromVolumeRecoveryPointResponse', createSnapshotFromVolumeRecoveryPointResponse_httpStatus - The response's http status code.

Response Lenses

createSnapshotFromVolumeRecoveryPointResponse_volumeRecoveryPointTime :: Lens' CreateSnapshotFromVolumeRecoveryPointResponse (Maybe Text)

The time the volume was created from the recovery point.

createSnapshotFromVolumeRecoveryPointResponse_volumeARN :: Lens' CreateSnapshotFromVolumeRecoveryPointResponse (Maybe Text)

The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

createSnapshotFromVolumeRecoveryPointResponse_snapshotId :: Lens' CreateSnapshotFromVolumeRecoveryPointResponse (Maybe Text)

The ID of the snapshot.

createSnapshotFromVolumeRecoveryPointResponse_httpStatus :: Lens' CreateSnapshotFromVolumeRecoveryPointResponse Int

The response's http status code.