Network.AWS.StorageGateway.CreateStorediSCSIVolume

Description

Creates a volume on a specified gateway. This operation is only supported in the stored volume gateway type.

The size of the volume to create is inferred from the disk size. You can choose to preserve existing data on the disk, create volume from an existing snapshot, or create an empty volume. If you choose to create an empty gateway volume, then any existing data on the disk is erased.

In the request, you must specify the gateway and the disk information on which you are creating the volume. In response, the gateway creates the volume and returns volume information such as the volume Amazon Resource Name (ARN), its size, and the iSCSI target ARN that initiators can use to connect to the volume target.

Creating a Request

data CreateStorediSCSIVolume

A JSON object containing one or more of the following fields:

CreateStorediSCSIVolumeInput$DiskId

CreateStorediSCSIVolumeInput$NetworkInterfaceId

CreateStorediSCSIVolumeInput$PreserveExistingData

CreateStorediSCSIVolumeInput$SnapshotId

CreateStorediSCSIVolumeInput$TargetName

See: newCreateStorediSCSIVolume smart constructor.

Constructors

CreateStorediSCSIVolume'

Fields

kmsKey :: Maybe Text

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

kmsEncrypted :: Maybe Bool

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

tags :: Maybe [Tag]

A list of up to 50 tags that can be assigned to a stored volume. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

snapshotId :: Maybe Text

The snapshot ID (e.g., "snap-1122aabb") of the snapshot to restore as the new stored volume. Specify this field if you want to create the iSCSI storage volume from a snapshot; otherwise, do not include this field. To list snapshots for your account use DescribeSnapshots in the Amazon Elastic Compute Cloud API Reference.

gatewayARN :: Text

diskId :: Text

The unique identifier for the gateway local disk that is configured as a stored volume. Use ListLocalDisks to list disk IDs for a gateway.

preserveExistingData :: Bool

Set to true if you want to preserve the data on the local disk. Otherwise, set to false to create an empty volume.

Valid Values: true | false

targetName :: Text

The name of the iSCSI target used by an initiator to connect to a volume and used as a suffix for the target ARN. For example, specifying TargetName as myvolume results in the target ARN of arn:aws:storagegateway:us-east-2:111122223333:gateway/sgw-12A3456B/target/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes on a gateway.

If you don't specify a value, Storage Gateway uses the value that was previously used for this volume as the new target name.

networkInterfaceId :: Text

The network interface of the gateway on which to expose the iSCSI target. Only IPv4 addresses are accepted. Use DescribeGatewayInformation to get a list of the network interfaces available on a gateway.

Valid Values: A valid IP address.

newCreateStorediSCSIVolume

Arguments

:: Text
$sel:gatewayARN:CreateStorediSCSIVolume'

-> Text
$sel:diskId:CreateStorediSCSIVolume'

-> Bool
$sel:preserveExistingData:CreateStorediSCSIVolume'

-> Text
$sel:targetName:CreateStorediSCSIVolume'

-> Text
$sel:networkInterfaceId:CreateStorediSCSIVolume'

-> CreateStorediSCSIVolume

Create a value of CreateStorediSCSIVolume with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKey:CreateStorediSCSIVolume', createStorediSCSIVolume_kmsKey - The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

$sel:kmsEncrypted:CreateStorediSCSIVolume', createStorediSCSIVolume_kmsEncrypted - Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

$sel:tags:CreateStorediSCSIVolume', createStorediSCSIVolume_tags - A list of up to 50 tags that can be assigned to a stored volume. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

$sel:snapshotId:CreateStorediSCSIVolume', createStorediSCSIVolume_snapshotId - The snapshot ID (e.g., "snap-1122aabb") of the snapshot to restore as the new stored volume. Specify this field if you want to create the iSCSI storage volume from a snapshot; otherwise, do not include this field. To list snapshots for your account use DescribeSnapshots in the Amazon Elastic Compute Cloud API Reference.

$sel:gatewayARN:CreateStorediSCSIVolume', createStorediSCSIVolume_gatewayARN - Undocumented member.

$sel:diskId:CreateStorediSCSIVolume', createStorediSCSIVolume_diskId - The unique identifier for the gateway local disk that is configured as a stored volume. Use ListLocalDisks to list disk IDs for a gateway.

$sel:preserveExistingData:CreateStorediSCSIVolume', createStorediSCSIVolume_preserveExistingData - Set to true if you want to preserve the data on the local disk. Otherwise, set to false to create an empty volume.

Valid Values: true | false

$sel:targetName:CreateStorediSCSIVolume', createStorediSCSIVolume_targetName - The name of the iSCSI target used by an initiator to connect to a volume and used as a suffix for the target ARN. For example, specifying TargetName as myvolume results in the target ARN of arn:aws:storagegateway:us-east-2:111122223333:gateway/sgw-12A3456B/target/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes on a gateway.

If you don't specify a value, Storage Gateway uses the value that was previously used for this volume as the new target name.

$sel:networkInterfaceId:CreateStorediSCSIVolume', createStorediSCSIVolume_networkInterfaceId - The network interface of the gateway on which to expose the iSCSI target. Only IPv4 addresses are accepted. Use DescribeGatewayInformation to get a list of the network interfaces available on a gateway.

Valid Values: A valid IP address.

Request Lenses

createStorediSCSIVolume_kmsKey :: Lens' CreateStorediSCSIVolume (Maybe Text)

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

createStorediSCSIVolume_kmsEncrypted :: Lens' CreateStorediSCSIVolume (Maybe Bool)

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

createStorediSCSIVolume_tags :: Lens' CreateStorediSCSIVolume (Maybe [Tag])

A list of up to 50 tags that can be assigned to a stored volume. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

createStorediSCSIVolume_snapshotId :: Lens' CreateStorediSCSIVolume (Maybe Text)

The snapshot ID (e.g., "snap-1122aabb") of the snapshot to restore as the new stored volume. Specify this field if you want to create the iSCSI storage volume from a snapshot; otherwise, do not include this field. To list snapshots for your account use DescribeSnapshots in the Amazon Elastic Compute Cloud API Reference.

createStorediSCSIVolume_gatewayARN :: Lens' CreateStorediSCSIVolume Text

Undocumented member.

createStorediSCSIVolume_diskId :: Lens' CreateStorediSCSIVolume Text

The unique identifier for the gateway local disk that is configured as a stored volume. Use ListLocalDisks to list disk IDs for a gateway.

createStorediSCSIVolume_preserveExistingData :: Lens' CreateStorediSCSIVolume Bool

Set to true if you want to preserve the data on the local disk. Otherwise, set to false to create an empty volume.

Valid Values: true | false

createStorediSCSIVolume_targetName :: Lens' CreateStorediSCSIVolume Text

The name of the iSCSI target used by an initiator to connect to a volume and used as a suffix for the target ARN. For example, specifying TargetName as myvolume results in the target ARN of arn:aws:storagegateway:us-east-2:111122223333:gateway/sgw-12A3456B/target/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes on a gateway.

If you don't specify a value, Storage Gateway uses the value that was previously used for this volume as the new target name.

createStorediSCSIVolume_networkInterfaceId :: Lens' CreateStorediSCSIVolume Text

The network interface of the gateway on which to expose the iSCSI target. Only IPv4 addresses are accepted. Use DescribeGatewayInformation to get a list of the network interfaces available on a gateway.

Valid Values: A valid IP address.

Destructuring the Response

data CreateStorediSCSIVolumeResponse

A JSON object containing the following fields:

See: newCreateStorediSCSIVolumeResponse smart constructor.

Constructors

CreateStorediSCSIVolumeResponse'

Fields

targetARN :: Maybe Text

The Amazon Resource Name (ARN) of the volume target, which includes the iSCSI name that initiators can use to connect to the target.

volumeARN :: Maybe Text

The Amazon Resource Name (ARN) of the configured volume.

volumeSizeInBytes :: Maybe Integer

The size of the volume in bytes.

httpStatus :: Int

The response's http status code.

newCreateStorediSCSIVolumeResponse

Arguments

:: Int
$sel:httpStatus:CreateStorediSCSIVolumeResponse'

-> CreateStorediSCSIVolumeResponse

Create a value of CreateStorediSCSIVolumeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetARN:CreateStorediSCSIVolumeResponse', createStorediSCSIVolumeResponse_targetARN - The Amazon Resource Name (ARN) of the volume target, which includes the iSCSI name that initiators can use to connect to the target.

$sel:volumeARN:CreateStorediSCSIVolumeResponse', createStorediSCSIVolumeResponse_volumeARN - The Amazon Resource Name (ARN) of the configured volume.

$sel:volumeSizeInBytes:CreateStorediSCSIVolumeResponse', createStorediSCSIVolumeResponse_volumeSizeInBytes - The size of the volume in bytes.

$sel:httpStatus:CreateStorediSCSIVolumeResponse', createStorediSCSIVolumeResponse_httpStatus - The response's http status code.

Response Lenses

createStorediSCSIVolumeResponse_targetARN :: Lens' CreateStorediSCSIVolumeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the volume target, which includes the iSCSI name that initiators can use to connect to the target.

createStorediSCSIVolumeResponse_volumeARN :: Lens' CreateStorediSCSIVolumeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the configured volume.

createStorediSCSIVolumeResponse_volumeSizeInBytes :: Lens' CreateStorediSCSIVolumeResponse (Maybe Integer)

The size of the volume in bytes.

createStorediSCSIVolumeResponse_httpStatus :: Lens' CreateStorediSCSIVolumeResponse Int

The response's http status code.