Network.AWS.StorageGateway.CreateTapes

Description

Creates one or more virtual tapes. You write data to the virtual tapes and then archive the tapes. This operation is only supported in the tape gateway type.

Cache storage must be allocated to the gateway before you can create virtual tapes. Use the AddCache operation to add cache storage to a gateway.

Creating a Request

data CreateTapes

CreateTapesInput

See: newCreateTapes smart constructor.

Constructors

CreateTapes'

Fields

kmsKey :: Maybe Text

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

kmsEncrypted :: Maybe Bool

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

poolId :: Maybe Text

The ID of the pool that you want to add your tape to for archiving. The tape in this pool is archived in the S3 storage class that is associated with the pool. When you use your backup application to eject the tape, the tape is archived directly into the storage class (S3 Glacier or S3 Glacier Deep Archive) that corresponds to the pool.

Valid Values: GLACIER | DEEP_ARCHIVE

worm :: Maybe Bool

Set to TRUE if the tape you are creating is to be configured as a write-once-read-many (WORM) tape.

tags :: Maybe [Tag]

A list of up to 50 tags that can be assigned to a virtual tape. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

gatewayARN :: Text

The unique Amazon Resource Name (ARN) that represents the gateway to associate the virtual tapes with. Use the ListGateways operation to return a list of gateways for your account and Region.

tapeSizeInBytes :: Integer

The size, in bytes, of the virtual tapes that you want to create.

The size must be aligned by gigabyte (1024*1024*1024 bytes).

clientToken :: Text

A unique identifier that you use to retry a request. If you retry a request, use the same ClientToken you specified in the initial request.

Using the same ClientToken prevents creating the tape multiple times.

numTapesToCreate :: Natural

The number of virtual tapes that you want to create.

tapeBarcodePrefix :: Text

A prefix that you append to the barcode of the virtual tape you are creating. This prefix makes the barcode unique.

The prefix must be 1-4 characters in length and must be one of the uppercase letters from A to Z.

newCreateTapes

Arguments

:: Text
$sel:gatewayARN:CreateTapes'

-> Integer
$sel:tapeSizeInBytes:CreateTapes'

-> Text
$sel:clientToken:CreateTapes'

-> Natural
$sel:numTapesToCreate:CreateTapes'

-> Text
$sel:tapeBarcodePrefix:CreateTapes'

-> CreateTapes

Create a value of CreateTapes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKey:CreateTapes', createTapes_kmsKey - The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

$sel:kmsEncrypted:CreateTapes', createTapes_kmsEncrypted - Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

$sel:poolId:CreateTapes', createTapes_poolId - The ID of the pool that you want to add your tape to for archiving. The tape in this pool is archived in the S3 storage class that is associated with the pool. When you use your backup application to eject the tape, the tape is archived directly into the storage class (S3 Glacier or S3 Glacier Deep Archive) that corresponds to the pool.

Valid Values: GLACIER | DEEP_ARCHIVE

$sel:worm:CreateTapes', createTapes_worm - Set to TRUE if the tape you are creating is to be configured as a write-once-read-many (WORM) tape.

$sel:tags:CreateTapes', createTapes_tags - A list of up to 50 tags that can be assigned to a virtual tape. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

$sel:gatewayARN:CreateTapes', createTapes_gatewayARN - The unique Amazon Resource Name (ARN) that represents the gateway to associate the virtual tapes with. Use the ListGateways operation to return a list of gateways for your account and Region.

$sel:tapeSizeInBytes:CreateTapes', createTapes_tapeSizeInBytes - The size, in bytes, of the virtual tapes that you want to create.

The size must be aligned by gigabyte (1024*1024*1024 bytes).

$sel:clientToken:CreateTapes', createTapes_clientToken - A unique identifier that you use to retry a request. If you retry a request, use the same ClientToken you specified in the initial request.

Using the same ClientToken prevents creating the tape multiple times.

$sel:numTapesToCreate:CreateTapes', createTapes_numTapesToCreate - The number of virtual tapes that you want to create.

$sel:tapeBarcodePrefix:CreateTapes', createTapes_tapeBarcodePrefix - A prefix that you append to the barcode of the virtual tape you are creating. This prefix makes the barcode unique.

The prefix must be 1-4 characters in length and must be one of the uppercase letters from A to Z.

Request Lenses

createTapes_kmsKey :: Lens' CreateTapes (Maybe Text)

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

createTapes_kmsEncrypted :: Lens' CreateTapes (Maybe Bool)

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

createTapes_poolId :: Lens' CreateTapes (Maybe Text)

The ID of the pool that you want to add your tape to for archiving. The tape in this pool is archived in the S3 storage class that is associated with the pool. When you use your backup application to eject the tape, the tape is archived directly into the storage class (S3 Glacier or S3 Glacier Deep Archive) that corresponds to the pool.

Valid Values: GLACIER | DEEP_ARCHIVE

createTapes_worm :: Lens' CreateTapes (Maybe Bool)

Set to TRUE if the tape you are creating is to be configured as a write-once-read-many (WORM) tape.

createTapes_tags :: Lens' CreateTapes (Maybe [Tag])

A list of up to 50 tags that can be assigned to a virtual tape. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

createTapes_gatewayARN :: Lens' CreateTapes Text

The unique Amazon Resource Name (ARN) that represents the gateway to associate the virtual tapes with. Use the ListGateways operation to return a list of gateways for your account and Region.

createTapes_tapeSizeInBytes :: Lens' CreateTapes Integer

The size, in bytes, of the virtual tapes that you want to create.

The size must be aligned by gigabyte (1024*1024*1024 bytes).

createTapes_clientToken :: Lens' CreateTapes Text

A unique identifier that you use to retry a request. If you retry a request, use the same ClientToken you specified in the initial request.

Using the same ClientToken prevents creating the tape multiple times.

createTapes_numTapesToCreate :: Lens' CreateTapes Natural

The number of virtual tapes that you want to create.

createTapes_tapeBarcodePrefix :: Lens' CreateTapes Text

A prefix that you append to the barcode of the virtual tape you are creating. This prefix makes the barcode unique.

The prefix must be 1-4 characters in length and must be one of the uppercase letters from A to Z.

Destructuring the Response

data CreateTapesResponse

CreateTapeOutput

See: newCreateTapesResponse smart constructor.

Constructors

CreateTapesResponse'

Fields

tapeARNs :: Maybe [Text]

A list of unique Amazon Resource Names (ARNs) that represents the virtual tapes that were created.

httpStatus :: Int

The response's http status code.

newCreateTapesResponse

Arguments

:: Int
$sel:httpStatus:CreateTapesResponse'

-> CreateTapesResponse

Create a value of CreateTapesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tapeARNs:CreateTapesResponse', createTapesResponse_tapeARNs - A list of unique Amazon Resource Names (ARNs) that represents the virtual tapes that were created.

$sel:httpStatus:CreateTapesResponse', createTapesResponse_httpStatus - The response's http status code.

Response Lenses

createTapesResponse_tapeARNs :: Lens' CreateTapesResponse (Maybe [Text])

A list of unique Amazon Resource Names (ARNs) that represents the virtual tapes that were created.

createTapesResponse_httpStatus :: Lens' CreateTapesResponse Int

The response's http status code.