Network.AWS.StorageGateway.DeleteChapCredentials

Description

Deletes Challenge-Handshake Authentication Protocol (CHAP) credentials for a specified iSCSI target and initiator pair. This operation is supported in volume and tape gateway types.

Creating a Request

data DeleteChapCredentials

A JSON object containing one or more of the following fields:

DeleteChapCredentialsInput$InitiatorName

DeleteChapCredentialsInput$TargetARN

See: newDeleteChapCredentials smart constructor.

Constructors

DeleteChapCredentials'

Fields

targetARN :: Text

The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

initiatorName :: Text

The iSCSI initiator that connects to the target.

newDeleteChapCredentials

Arguments

:: Text
$sel:targetARN:DeleteChapCredentials'

-> Text
$sel:initiatorName:DeleteChapCredentials'

-> DeleteChapCredentials

Create a value of DeleteChapCredentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetARN:DeleteChapCredentials', deleteChapCredentials_targetARN - The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

$sel:initiatorName:DeleteChapCredentials', deleteChapCredentials_initiatorName - The iSCSI initiator that connects to the target.

Request Lenses

deleteChapCredentials_targetARN :: Lens' DeleteChapCredentials Text

The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

deleteChapCredentials_initiatorName :: Lens' DeleteChapCredentials Text

The iSCSI initiator that connects to the target.

Destructuring the Response

data DeleteChapCredentialsResponse

A JSON object containing the following fields:

See: newDeleteChapCredentialsResponse smart constructor.

Constructors

DeleteChapCredentialsResponse'

Fields

targetARN :: Maybe Text

The Amazon Resource Name (ARN) of the target.

initiatorName :: Maybe Text

The iSCSI initiator that connects to the target.

httpStatus :: Int

The response's http status code.

newDeleteChapCredentialsResponse

Arguments

:: Int
$sel:httpStatus:DeleteChapCredentialsResponse'

-> DeleteChapCredentialsResponse

Create a value of DeleteChapCredentialsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetARN:DeleteChapCredentials', deleteChapCredentialsResponse_targetARN - The Amazon Resource Name (ARN) of the target.

$sel:initiatorName:DeleteChapCredentials', deleteChapCredentialsResponse_initiatorName - The iSCSI initiator that connects to the target.

$sel:httpStatus:DeleteChapCredentialsResponse', deleteChapCredentialsResponse_httpStatus - The response's http status code.

Response Lenses

deleteChapCredentialsResponse_targetARN :: Lens' DeleteChapCredentialsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the target.

deleteChapCredentialsResponse_initiatorName :: Lens' DeleteChapCredentialsResponse (Maybe Text)

The iSCSI initiator that connects to the target.

deleteChapCredentialsResponse_httpStatus :: Lens' DeleteChapCredentialsResponse Int

The response's http status code.