Network.AWS.StorageGateway.DeleteVolume

Description

Deletes the specified storage volume that you previously created using the CreateCachediSCSIVolume or CreateStorediSCSIVolume API. This operation is only supported in the cached volume and stored volume types. For stored volume gateways, the local disk that was configured as the storage volume is not deleted. You can reuse the local disk to create another storage volume.

Before you delete a volume, make sure there are no iSCSI connections to the volume you are deleting. You should also make sure there is no snapshot in progress. You can use the Amazon Elastic Compute Cloud (Amazon EC2) API to query snapshots on the volume you are deleting and check the snapshot status. For more information, go to DescribeSnapshots in the Amazon Elastic Compute Cloud API Reference.

In the request, you must provide the Amazon Resource Name (ARN) of the storage volume you want to delete.

Creating a Request

data DeleteVolume

A JSON object containing the DeleteVolumeInput$VolumeARN to delete.

See: newDeleteVolume smart constructor.

Constructors

DeleteVolume'

Fields

volumeARN :: Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

newDeleteVolume

Arguments

:: Text
$sel:volumeARN:DeleteVolume'

-> DeleteVolume

Create a value of DeleteVolume with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:volumeARN:DeleteVolume', deleteVolume_volumeARN - The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

Request Lenses

deleteVolume_volumeARN :: Lens' DeleteVolume Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

Destructuring the Response

data DeleteVolumeResponse

A JSON object containing the Amazon Resource Name (ARN) of the storage volume that was deleted.

See: newDeleteVolumeResponse smart constructor.

Constructors

DeleteVolumeResponse'

Fields

volumeARN :: Maybe Text

The Amazon Resource Name (ARN) of the storage volume that was deleted. It is the same ARN you provided in the request.

httpStatus :: Int

The response's http status code.

newDeleteVolumeResponse

Arguments

:: Int
$sel:httpStatus:DeleteVolumeResponse'

-> DeleteVolumeResponse

Create a value of DeleteVolumeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:volumeARN:DeleteVolume', deleteVolumeResponse_volumeARN - The Amazon Resource Name (ARN) of the storage volume that was deleted. It is the same ARN you provided in the request.

$sel:httpStatus:DeleteVolumeResponse', deleteVolumeResponse_httpStatus - The response's http status code.

Response Lenses

deleteVolumeResponse_volumeARN :: Lens' DeleteVolumeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the storage volume that was deleted. It is the same ARN you provided in the request.

deleteVolumeResponse_httpStatus :: Lens' DeleteVolumeResponse Int

The response's http status code.