Network.AWS.StorageGateway.DescribeBandwidthRateLimit

Description

Returns the bandwidth rate limits of a gateway. By default, these limits are not set, which means no bandwidth rate limiting is in effect. This operation is supported for the stored volume, cached volume, and tape gateway types.

This operation only returns a value for a bandwidth rate limit only if the limit is set. If no limits are set for the gateway, then this operation returns only the gateway ARN in the response body. To specify which gateway to describe, use the Amazon Resource Name (ARN) of the gateway in your request.

Creating a Request

data DescribeBandwidthRateLimit

A JSON object containing the Amazon Resource Name (ARN) of the gateway.

See: newDescribeBandwidthRateLimit smart constructor.

Constructors

DescribeBandwidthRateLimit'

Fields

gatewayARN :: Text

newDescribeBandwidthRateLimit

Arguments

:: Text
$sel:gatewayARN:DescribeBandwidthRateLimit'

-> DescribeBandwidthRateLimit

Create a value of DescribeBandwidthRateLimit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:DescribeBandwidthRateLimit', describeBandwidthRateLimit_gatewayARN - Undocumented member.

Request Lenses

describeBandwidthRateLimit_gatewayARN :: Lens' DescribeBandwidthRateLimit Text

Undocumented member.

Destructuring the Response

data DescribeBandwidthRateLimitResponse

A JSON object containing the following fields:

See: newDescribeBandwidthRateLimitResponse smart constructor.

Constructors

DescribeBandwidthRateLimitResponse'

Fields

gatewayARN :: Maybe Text

averageUploadRateLimitInBitsPerSec :: Maybe Natural

The average upload bandwidth rate limit in bits per second. This field does not appear in the response if the upload rate limit is not set.

averageDownloadRateLimitInBitsPerSec :: Maybe Natural

The average download bandwidth rate limit in bits per second. This field does not appear in the response if the download rate limit is not set.

httpStatus :: Int

The response's http status code.

newDescribeBandwidthRateLimitResponse

Arguments

:: Int
$sel:httpStatus:DescribeBandwidthRateLimitResponse'

-> DescribeBandwidthRateLimitResponse

Create a value of DescribeBandwidthRateLimitResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:DescribeBandwidthRateLimit', describeBandwidthRateLimitResponse_gatewayARN - Undocumented member.

$sel:averageUploadRateLimitInBitsPerSec:DescribeBandwidthRateLimitResponse', describeBandwidthRateLimitResponse_averageUploadRateLimitInBitsPerSec - The average upload bandwidth rate limit in bits per second. This field does not appear in the response if the upload rate limit is not set.

$sel:averageDownloadRateLimitInBitsPerSec:DescribeBandwidthRateLimitResponse', describeBandwidthRateLimitResponse_averageDownloadRateLimitInBitsPerSec - The average download bandwidth rate limit in bits per second. This field does not appear in the response if the download rate limit is not set.

$sel:httpStatus:DescribeBandwidthRateLimitResponse', describeBandwidthRateLimitResponse_httpStatus - The response's http status code.

Response Lenses

describeBandwidthRateLimitResponse_gatewayARN :: Lens' DescribeBandwidthRateLimitResponse (Maybe Text)

Undocumented member.

describeBandwidthRateLimitResponse_averageUploadRateLimitInBitsPerSec :: Lens' DescribeBandwidthRateLimitResponse (Maybe Natural)

The average upload bandwidth rate limit in bits per second. This field does not appear in the response if the upload rate limit is not set.

describeBandwidthRateLimitResponse_averageDownloadRateLimitInBitsPerSec :: Lens' DescribeBandwidthRateLimitResponse (Maybe Natural)

The average download bandwidth rate limit in bits per second. This field does not appear in the response if the download rate limit is not set.

describeBandwidthRateLimitResponse_httpStatus :: Lens' DescribeBandwidthRateLimitResponse Int

The response's http status code.