Network.AWS.StorageGateway.DescribeBandwidthRateLimitSchedule

Description

Returns information about the bandwidth rate limit schedule of a gateway. By default, gateways do not have bandwidth rate limit schedules, which means no bandwidth rate limiting is in effect. This operation is supported only in the volume and tape gateway types.

This operation returns information about a gateway's bandwidth rate limit schedule. A bandwidth rate limit schedule consists of one or more bandwidth rate limit intervals. A bandwidth rate limit interval defines a period of time on one or more days of the week, during which bandwidth rate limits are specified for uploading, downloading, or both.

A bandwidth rate limit interval consists of one or more days of the week, a start hour and minute, an ending hour and minute, and bandwidth rate limits for uploading and downloading

If no bandwidth rate limit schedule intervals are set for the gateway, this operation returns an empty response. To specify which gateway to describe, use the Amazon Resource Name (ARN) of the gateway in your request.

Creating a Request

data DescribeBandwidthRateLimitSchedule

See: newDescribeBandwidthRateLimitSchedule smart constructor.

Constructors

DescribeBandwidthRateLimitSchedule'

Fields

gatewayARN :: Text

newDescribeBandwidthRateLimitSchedule

Arguments

:: Text
$sel:gatewayARN:DescribeBandwidthRateLimitSchedule'

-> DescribeBandwidthRateLimitSchedule

Create a value of DescribeBandwidthRateLimitSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:DescribeBandwidthRateLimitSchedule', describeBandwidthRateLimitSchedule_gatewayARN - Undocumented member.

Request Lenses

describeBandwidthRateLimitSchedule_gatewayARN :: Lens' DescribeBandwidthRateLimitSchedule Text

Undocumented member.

Destructuring the Response

data DescribeBandwidthRateLimitScheduleResponse

See: newDescribeBandwidthRateLimitScheduleResponse smart constructor.

Constructors

DescribeBandwidthRateLimitScheduleResponse'

Fields

gatewayARN :: Maybe Text

bandwidthRateLimitIntervals :: Maybe [BandwidthRateLimitInterval]

An array that contains the bandwidth rate limit intervals for a tape or volume gateway.

httpStatus :: Int

The response's http status code.

newDescribeBandwidthRateLimitScheduleResponse

Arguments

:: Int
$sel:httpStatus:DescribeBandwidthRateLimitScheduleResponse'

-> DescribeBandwidthRateLimitScheduleResponse

Create a value of DescribeBandwidthRateLimitScheduleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:DescribeBandwidthRateLimitSchedule', describeBandwidthRateLimitScheduleResponse_gatewayARN - Undocumented member.

$sel:bandwidthRateLimitIntervals:DescribeBandwidthRateLimitScheduleResponse', describeBandwidthRateLimitScheduleResponse_bandwidthRateLimitIntervals - An array that contains the bandwidth rate limit intervals for a tape or volume gateway.

$sel:httpStatus:DescribeBandwidthRateLimitScheduleResponse', describeBandwidthRateLimitScheduleResponse_httpStatus - The response's http status code.

Response Lenses

describeBandwidthRateLimitScheduleResponse_gatewayARN :: Lens' DescribeBandwidthRateLimitScheduleResponse (Maybe Text)

Undocumented member.

describeBandwidthRateLimitScheduleResponse_bandwidthRateLimitIntervals :: Lens' DescribeBandwidthRateLimitScheduleResponse (Maybe [BandwidthRateLimitInterval])

An array that contains the bandwidth rate limit intervals for a tape or volume gateway.

describeBandwidthRateLimitScheduleResponse_httpStatus :: Lens' DescribeBandwidthRateLimitScheduleResponse Int

The response's http status code.