Network.AWS.StorageGateway.DescribeChapCredentials

Description

Returns an array of Challenge-Handshake Authentication Protocol (CHAP) credentials information for a specified iSCSI target, one for each target-initiator pair. This operation is supported in the volume and tape gateway types.

Creating a Request

data DescribeChapCredentials

A JSON object containing the Amazon Resource Name (ARN) of the iSCSI volume target.

See: newDescribeChapCredentials smart constructor.

Constructors

DescribeChapCredentials'

Fields

targetARN :: Text

The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

newDescribeChapCredentials

Arguments

:: Text
$sel:targetARN:DescribeChapCredentials'

-> DescribeChapCredentials

Create a value of DescribeChapCredentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetARN:DescribeChapCredentials', describeChapCredentials_targetARN - The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

Request Lenses

describeChapCredentials_targetARN :: Lens' DescribeChapCredentials Text

The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

Destructuring the Response

data DescribeChapCredentialsResponse

A JSON object containing the following fields:

See: newDescribeChapCredentialsResponse smart constructor.

Constructors

DescribeChapCredentialsResponse'

Fields

chapCredentials :: Maybe [ChapInfo]

An array of ChapInfo objects that represent CHAP credentials. Each object in the array contains CHAP credential information for one target-initiator pair. If no CHAP credentials are set, an empty array is returned. CHAP credential information is provided in a JSON object with the following fields:

InitiatorName: The iSCSI initiator that connects to the target.

SecretToAuthenticateInitiator: The secret key that the initiator (for example, the Windows client) must provide to participate in mutual CHAP with the target.

SecretToAuthenticateTarget: The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g. Windows client).

TargetARN: The Amazon Resource Name (ARN) of the storage volume.

httpStatus :: Int

The response's http status code.

newDescribeChapCredentialsResponse

Arguments

:: Int
$sel:httpStatus:DescribeChapCredentialsResponse'

-> DescribeChapCredentialsResponse

Create a value of DescribeChapCredentialsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:chapCredentials:DescribeChapCredentialsResponse', describeChapCredentialsResponse_chapCredentials - An array of ChapInfo objects that represent CHAP credentials. Each object in the array contains CHAP credential information for one target-initiator pair. If no CHAP credentials are set, an empty array is returned. CHAP credential information is provided in a JSON object with the following fields:

InitiatorName: The iSCSI initiator that connects to the target.

SecretToAuthenticateInitiator: The secret key that the initiator (for example, the Windows client) must provide to participate in mutual CHAP with the target.

SecretToAuthenticateTarget: The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g. Windows client).

TargetARN: The Amazon Resource Name (ARN) of the storage volume.

$sel:httpStatus:DescribeChapCredentialsResponse', describeChapCredentialsResponse_httpStatus - The response's http status code.

Response Lenses

describeChapCredentialsResponse_chapCredentials :: Lens' DescribeChapCredentialsResponse (Maybe [ChapInfo])

An array of ChapInfo objects that represent CHAP credentials. Each object in the array contains CHAP credential information for one target-initiator pair. If no CHAP credentials are set, an empty array is returned. CHAP credential information is provided in a JSON object with the following fields:

InitiatorName: The iSCSI initiator that connects to the target.

SecretToAuthenticateInitiator: The secret key that the initiator (for example, the Windows client) must provide to participate in mutual CHAP with the target.

SecretToAuthenticateTarget: The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g. Windows client).

TargetARN: The Amazon Resource Name (ARN) of the storage volume.

describeChapCredentialsResponse_httpStatus :: Lens' DescribeChapCredentialsResponse Int

The response's http status code.