Network.AWS.StorageGateway.DescribeGatewayInformation

Description

Returns metadata about a gateway such as its name, network interfaces, configured time zone, and the state (whether the gateway is running or not). To specify which gateway to describe, use the Amazon Resource Name (ARN) of the gateway in your request.

Creating a Request

data DescribeGatewayInformation

A JSON object containing the ID of the gateway.

See: newDescribeGatewayInformation smart constructor.

Constructors

DescribeGatewayInformation'

Fields

gatewayARN :: Text

newDescribeGatewayInformation

Arguments

:: Text
$sel:gatewayARN:DescribeGatewayInformation'

-> DescribeGatewayInformation

Create a value of DescribeGatewayInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:DescribeGatewayInformation', describeGatewayInformation_gatewayARN - Undocumented member.

Request Lenses

describeGatewayInformation_gatewayARN :: Lens' DescribeGatewayInformation Text

Undocumented member.

Destructuring the Response

data DescribeGatewayInformationResponse

A JSON object containing the following fields:

See: newDescribeGatewayInformationResponse smart constructor.

Constructors

DescribeGatewayInformationResponse'

Fields

gatewayState :: Maybe Text

A value that indicates the operating state of the gateway.

ec2InstanceRegion :: Maybe Text

The Region where the Amazon EC2 instance is located.

gatewayARN :: Maybe Text

gatewayNetworkInterfaces :: Maybe [NetworkInterface]

A NetworkInterface array that contains descriptions of the gateway network interfaces.

ec2InstanceId :: Maybe Text

The ID of the Amazon EC2 instance that was used to launch the gateway.

nextUpdateAvailabilityDate :: Maybe Text

The date on which an update to the gateway is available. This date is in the time zone of the gateway. If the gateway is not available for an update this field is not returned in the response.

endpointType :: Maybe Text

The type of endpoint for your gateway.

Valid Values: STANDARD | FIPS

deprecationDate :: Maybe Text

Date after which this gateway will not receive software updates for new features and bug fixes.

lastSoftwareUpdate :: Maybe Text

The date on which the last software update was applied to the gateway. If the gateway has never been updated, this field does not return a value in the response.

gatewayCapacity :: Maybe GatewayCapacity

Specifies the size of the gateway's metadata cache.

supportedGatewayCapacities :: Maybe [GatewayCapacity]

A list of the metadata cache sizes that the gateway can support based on its current hardware specifications.

gatewayName :: Maybe Text

The name you configured for your gateway.

gatewayId :: Maybe Text

The unique identifier assigned to your gateway during activation. This ID becomes part of the gateway Amazon Resource Name (ARN), which you use as input for other operations.

hostEnvironment :: Maybe HostEnvironment

The type of hypervisor environment used by the host.

gatewayType :: Maybe Text

The type of the gateway.

gatewayTimezone :: Maybe Text

A value that indicates the time zone configured for the gateway.

softwareUpdatesEndDate :: Maybe Text

Date after which this gateway will not receive software updates for new features.

cloudWatchLogGroupARN :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon CloudWatch log group that is used to monitor events in the gateway.

vPCEndpoint :: Maybe Text

The configuration settings for the virtual private cloud (VPC) endpoint for your gateway.

tags :: Maybe [Tag]

A list of up to 50 tags assigned to the gateway, sorted alphabetically by key name. Each tag is a key-value pair. For a gateway with more than 10 tags assigned, you can view all tags using the ListTagsForResource API operation.

httpStatus :: Int

The response's http status code.

newDescribeGatewayInformationResponse

Arguments

:: Int
$sel:httpStatus:DescribeGatewayInformationResponse'

-> DescribeGatewayInformationResponse

Create a value of DescribeGatewayInformationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayState:DescribeGatewayInformationResponse', describeGatewayInformationResponse_gatewayState - A value that indicates the operating state of the gateway.

$sel:ec2InstanceRegion:DescribeGatewayInformationResponse', describeGatewayInformationResponse_ec2InstanceRegion - The Region where the Amazon EC2 instance is located.

$sel:gatewayARN:DescribeGatewayInformation', describeGatewayInformationResponse_gatewayARN - Undocumented member.

$sel:gatewayNetworkInterfaces:DescribeGatewayInformationResponse', describeGatewayInformationResponse_gatewayNetworkInterfaces - A NetworkInterface array that contains descriptions of the gateway network interfaces.

$sel:ec2InstanceId:DescribeGatewayInformationResponse', describeGatewayInformationResponse_ec2InstanceId - The ID of the Amazon EC2 instance that was used to launch the gateway.

$sel:nextUpdateAvailabilityDate:DescribeGatewayInformationResponse', describeGatewayInformationResponse_nextUpdateAvailabilityDate - The date on which an update to the gateway is available. This date is in the time zone of the gateway. If the gateway is not available for an update this field is not returned in the response.

$sel:endpointType:DescribeGatewayInformationResponse', describeGatewayInformationResponse_endpointType - The type of endpoint for your gateway.

Valid Values: STANDARD | FIPS

$sel:deprecationDate:DescribeGatewayInformationResponse', describeGatewayInformationResponse_deprecationDate - Date after which this gateway will not receive software updates for new features and bug fixes.

$sel:lastSoftwareUpdate:DescribeGatewayInformationResponse', describeGatewayInformationResponse_lastSoftwareUpdate - The date on which the last software update was applied to the gateway. If the gateway has never been updated, this field does not return a value in the response.

$sel:gatewayCapacity:DescribeGatewayInformationResponse', describeGatewayInformationResponse_gatewayCapacity - Specifies the size of the gateway's metadata cache.

$sel:supportedGatewayCapacities:DescribeGatewayInformationResponse', describeGatewayInformationResponse_supportedGatewayCapacities - A list of the metadata cache sizes that the gateway can support based on its current hardware specifications.

$sel:gatewayName:DescribeGatewayInformationResponse', describeGatewayInformationResponse_gatewayName - The name you configured for your gateway.

$sel:gatewayId:DescribeGatewayInformationResponse', describeGatewayInformationResponse_gatewayId - The unique identifier assigned to your gateway during activation. This ID becomes part of the gateway Amazon Resource Name (ARN), which you use as input for other operations.

$sel:hostEnvironment:DescribeGatewayInformationResponse', describeGatewayInformationResponse_hostEnvironment - The type of hypervisor environment used by the host.

$sel:gatewayType:DescribeGatewayInformationResponse', describeGatewayInformationResponse_gatewayType - The type of the gateway.

$sel:gatewayTimezone:DescribeGatewayInformationResponse', describeGatewayInformationResponse_gatewayTimezone - A value that indicates the time zone configured for the gateway.

$sel:softwareUpdatesEndDate:DescribeGatewayInformationResponse', describeGatewayInformationResponse_softwareUpdatesEndDate - Date after which this gateway will not receive software updates for new features.

$sel:cloudWatchLogGroupARN:DescribeGatewayInformationResponse', describeGatewayInformationResponse_cloudWatchLogGroupARN - The Amazon Resource Name (ARN) of the Amazon CloudWatch log group that is used to monitor events in the gateway.

$sel:vPCEndpoint:DescribeGatewayInformationResponse', describeGatewayInformationResponse_vPCEndpoint - The configuration settings for the virtual private cloud (VPC) endpoint for your gateway.

$sel:tags:DescribeGatewayInformationResponse', describeGatewayInformationResponse_tags - A list of up to 50 tags assigned to the gateway, sorted alphabetically by key name. Each tag is a key-value pair. For a gateway with more than 10 tags assigned, you can view all tags using the ListTagsForResource API operation.

$sel:httpStatus:DescribeGatewayInformationResponse', describeGatewayInformationResponse_httpStatus - The response's http status code.

Response Lenses

describeGatewayInformationResponse_gatewayState :: Lens' DescribeGatewayInformationResponse (Maybe Text)

A value that indicates the operating state of the gateway.

describeGatewayInformationResponse_ec2InstanceRegion :: Lens' DescribeGatewayInformationResponse (Maybe Text)

The Region where the Amazon EC2 instance is located.

describeGatewayInformationResponse_gatewayARN :: Lens' DescribeGatewayInformationResponse (Maybe Text)

Undocumented member.

describeGatewayInformationResponse_gatewayNetworkInterfaces :: Lens' DescribeGatewayInformationResponse (Maybe [NetworkInterface])

A NetworkInterface array that contains descriptions of the gateway network interfaces.

describeGatewayInformationResponse_ec2InstanceId :: Lens' DescribeGatewayInformationResponse (Maybe Text)

The ID of the Amazon EC2 instance that was used to launch the gateway.

describeGatewayInformationResponse_nextUpdateAvailabilityDate :: Lens' DescribeGatewayInformationResponse (Maybe Text)

The date on which an update to the gateway is available. This date is in the time zone of the gateway. If the gateway is not available for an update this field is not returned in the response.

describeGatewayInformationResponse_endpointType :: Lens' DescribeGatewayInformationResponse (Maybe Text)

The type of endpoint for your gateway.

Valid Values: STANDARD | FIPS

describeGatewayInformationResponse_deprecationDate :: Lens' DescribeGatewayInformationResponse (Maybe Text)

Date after which this gateway will not receive software updates for new features and bug fixes.

describeGatewayInformationResponse_lastSoftwareUpdate :: Lens' DescribeGatewayInformationResponse (Maybe Text)

The date on which the last software update was applied to the gateway. If the gateway has never been updated, this field does not return a value in the response.

describeGatewayInformationResponse_gatewayCapacity :: Lens' DescribeGatewayInformationResponse (Maybe GatewayCapacity)

Specifies the size of the gateway's metadata cache.

describeGatewayInformationResponse_supportedGatewayCapacities :: Lens' DescribeGatewayInformationResponse (Maybe [GatewayCapacity])

A list of the metadata cache sizes that the gateway can support based on its current hardware specifications.

describeGatewayInformationResponse_gatewayName :: Lens' DescribeGatewayInformationResponse (Maybe Text)

The name you configured for your gateway.

describeGatewayInformationResponse_gatewayId :: Lens' DescribeGatewayInformationResponse (Maybe Text)

The unique identifier assigned to your gateway during activation. This ID becomes part of the gateway Amazon Resource Name (ARN), which you use as input for other operations.

describeGatewayInformationResponse_hostEnvironment :: Lens' DescribeGatewayInformationResponse (Maybe HostEnvironment)

The type of hypervisor environment used by the host.

describeGatewayInformationResponse_gatewayType :: Lens' DescribeGatewayInformationResponse (Maybe Text)

The type of the gateway.

describeGatewayInformationResponse_gatewayTimezone :: Lens' DescribeGatewayInformationResponse (Maybe Text)

A value that indicates the time zone configured for the gateway.

describeGatewayInformationResponse_softwareUpdatesEndDate :: Lens' DescribeGatewayInformationResponse (Maybe Text)

Date after which this gateway will not receive software updates for new features.

describeGatewayInformationResponse_cloudWatchLogGroupARN :: Lens' DescribeGatewayInformationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon CloudWatch log group that is used to monitor events in the gateway.

describeGatewayInformationResponse_vPCEndpoint :: Lens' DescribeGatewayInformationResponse (Maybe Text)

The configuration settings for the virtual private cloud (VPC) endpoint for your gateway.

describeGatewayInformationResponse_tags :: Lens' DescribeGatewayInformationResponse (Maybe [Tag])

A list of up to 50 tags assigned to the gateway, sorted alphabetically by key name. Each tag is a key-value pair. For a gateway with more than 10 tags assigned, you can view all tags using the ListTagsForResource API operation.

describeGatewayInformationResponse_httpStatus :: Lens' DescribeGatewayInformationResponse Int

The response's http status code.