Network.AWS.StorageGateway.DescribeNFSFileShares

Description

Gets a description for one or more Network File System (NFS) file shares from an S3 File Gateway. This operation is only supported for S3 File Gateways.

Creating a Request

data DescribeNFSFileShares

DescribeNFSFileSharesInput

See: newDescribeNFSFileShares smart constructor.

Constructors

DescribeNFSFileShares'

Fields

fileShareARNList :: NonEmpty Text

An array containing the Amazon Resource Name (ARN) of each file share to be described.

newDescribeNFSFileShares

Arguments

:: NonEmpty Text
$sel:fileShareARNList:DescribeNFSFileShares'

-> DescribeNFSFileShares

Create a value of DescribeNFSFileShares with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fileShareARNList:DescribeNFSFileShares', describeNFSFileShares_fileShareARNList - An array containing the Amazon Resource Name (ARN) of each file share to be described.

Request Lenses

describeNFSFileShares_fileShareARNList :: Lens' DescribeNFSFileShares (NonEmpty Text)

An array containing the Amazon Resource Name (ARN) of each file share to be described.

Destructuring the Response

data DescribeNFSFileSharesResponse

DescribeNFSFileSharesOutput

See: newDescribeNFSFileSharesResponse smart constructor.

Constructors

DescribeNFSFileSharesResponse'

Fields

nFSFileShareInfoList :: Maybe [NFSFileShareInfo]

An array containing a description for each requested file share.

httpStatus :: Int

The response's http status code.

newDescribeNFSFileSharesResponse

Arguments

:: Int
$sel:httpStatus:DescribeNFSFileSharesResponse'

-> DescribeNFSFileSharesResponse

Create a value of DescribeNFSFileSharesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nFSFileShareInfoList:DescribeNFSFileSharesResponse', describeNFSFileSharesResponse_nFSFileShareInfoList - An array containing a description for each requested file share.

$sel:httpStatus:DescribeNFSFileSharesResponse', describeNFSFileSharesResponse_httpStatus - The response's http status code.

Response Lenses

describeNFSFileSharesResponse_nFSFileShareInfoList :: Lens' DescribeNFSFileSharesResponse (Maybe [NFSFileShareInfo])

An array containing a description for each requested file share.

describeNFSFileSharesResponse_httpStatus :: Lens' DescribeNFSFileSharesResponse Int

The response's http status code.