Network.AWS.StorageGateway.DescribeSMBSettings

Description

Gets a description of a Server Message Block (SMB) file share settings from a file gateway. This operation is only supported for file gateways.

Creating a Request

data DescribeSMBSettings

See: newDescribeSMBSettings smart constructor.

Constructors

DescribeSMBSettings'

Fields

gatewayARN :: Text

newDescribeSMBSettings

Arguments

:: Text
$sel:gatewayARN:DescribeSMBSettings'

-> DescribeSMBSettings

Create a value of DescribeSMBSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:DescribeSMBSettings', describeSMBSettings_gatewayARN - Undocumented member.

Request Lenses

describeSMBSettings_gatewayARN :: Lens' DescribeSMBSettings Text

Undocumented member.

Destructuring the Response

data DescribeSMBSettingsResponse

See: newDescribeSMBSettingsResponse smart constructor.

Constructors

DescribeSMBSettingsResponse'

Fields

gatewayARN :: Maybe Text

fileSharesVisible :: Maybe Bool

The shares on this gateway appear when listing shares. Only supported for S3 File Gateways.

activeDirectoryStatus :: Maybe ActiveDirectoryStatus

Indicates the status of a gateway that is a member of the Active Directory domain.

ACCESS_DENIED: Indicates that the JoinDomain operation failed due to an authentication error.

DETACHED: Indicates that gateway is not joined to a domain.

JOINED: Indicates that the gateway has successfully joined a domain.

JOINING: Indicates that a JoinDomain operation is in progress.

NETWORK_ERROR: Indicates that JoinDomain operation failed due to a network or connectivity error.

TIMEOUT: Indicates that the JoinDomain operation failed because the operation didn't complete within the allotted time.

UNKNOWN_ERROR: Indicates that the JoinDomain operation failed due to another type of error.

domainName :: Maybe Text

The name of the domain that the gateway is joined to.

sMBGuestPasswordSet :: Maybe Bool

This value is true if a password for the guest user smbguest is set, otherwise false. Only supported for S3 File Gateways.

Valid Values: true | false

sMBSecurityStrategy :: Maybe SMBSecurityStrategy

The type of security strategy that was specified for file gateway.

ClientSpecified: If you use this option, requests are established based on what is negotiated by the client. This option is recommended when you want to maximize compatibility across different clients in your environment. Only supported for S3 File Gateways.

MandatorySigning: If you use this option, file gateway only allows connections from SMBv2 or SMBv3 clients that have signing enabled. This option works with SMB clients on Microsoft Windows Vista, Windows Server 2008 or newer.

MandatoryEncryption: If you use this option, file gateway only allows connections from SMBv3 clients that have encryption enabled. This option is highly recommended for environments that handle sensitive data. This option works with SMB clients on Microsoft Windows 8, Windows Server 2012 or newer.

httpStatus :: Int

The response's http status code.

newDescribeSMBSettingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeSMBSettingsResponse'

-> DescribeSMBSettingsResponse

Create a value of DescribeSMBSettingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:DescribeSMBSettings', describeSMBSettingsResponse_gatewayARN - Undocumented member.

$sel:fileSharesVisible:DescribeSMBSettingsResponse', describeSMBSettingsResponse_fileSharesVisible - The shares on this gateway appear when listing shares. Only supported for S3 File Gateways.

$sel:activeDirectoryStatus:DescribeSMBSettingsResponse', describeSMBSettingsResponse_activeDirectoryStatus - Indicates the status of a gateway that is a member of the Active Directory domain.

ACCESS_DENIED: Indicates that the JoinDomain operation failed due to an authentication error.

DETACHED: Indicates that gateway is not joined to a domain.

JOINED: Indicates that the gateway has successfully joined a domain.

JOINING: Indicates that a JoinDomain operation is in progress.

NETWORK_ERROR: Indicates that JoinDomain operation failed due to a network or connectivity error.

TIMEOUT: Indicates that the JoinDomain operation failed because the operation didn't complete within the allotted time.

UNKNOWN_ERROR: Indicates that the JoinDomain operation failed due to another type of error.

$sel:domainName:DescribeSMBSettingsResponse', describeSMBSettingsResponse_domainName - The name of the domain that the gateway is joined to.

$sel:sMBGuestPasswordSet:DescribeSMBSettingsResponse', describeSMBSettingsResponse_sMBGuestPasswordSet - This value is true if a password for the guest user smbguest is set, otherwise false. Only supported for S3 File Gateways.

Valid Values: true | false

$sel:sMBSecurityStrategy:DescribeSMBSettingsResponse', describeSMBSettingsResponse_sMBSecurityStrategy - The type of security strategy that was specified for file gateway.

ClientSpecified: If you use this option, requests are established based on what is negotiated by the client. This option is recommended when you want to maximize compatibility across different clients in your environment. Only supported for S3 File Gateways.

MandatorySigning: If you use this option, file gateway only allows connections from SMBv2 or SMBv3 clients that have signing enabled. This option works with SMB clients on Microsoft Windows Vista, Windows Server 2008 or newer.

MandatoryEncryption: If you use this option, file gateway only allows connections from SMBv3 clients that have encryption enabled. This option is highly recommended for environments that handle sensitive data. This option works with SMB clients on Microsoft Windows 8, Windows Server 2012 or newer.

$sel:httpStatus:DescribeSMBSettingsResponse', describeSMBSettingsResponse_httpStatus - The response's http status code.

Response Lenses

describeSMBSettingsResponse_gatewayARN :: Lens' DescribeSMBSettingsResponse (Maybe Text)

Undocumented member.

describeSMBSettingsResponse_fileSharesVisible :: Lens' DescribeSMBSettingsResponse (Maybe Bool)

The shares on this gateway appear when listing shares. Only supported for S3 File Gateways.

describeSMBSettingsResponse_activeDirectoryStatus :: Lens' DescribeSMBSettingsResponse (Maybe ActiveDirectoryStatus)

Indicates the status of a gateway that is a member of the Active Directory domain.

ACCESS_DENIED: Indicates that the JoinDomain operation failed due to an authentication error.

DETACHED: Indicates that gateway is not joined to a domain.

JOINED: Indicates that the gateway has successfully joined a domain.

JOINING: Indicates that a JoinDomain operation is in progress.

NETWORK_ERROR: Indicates that JoinDomain operation failed due to a network or connectivity error.

TIMEOUT: Indicates that the JoinDomain operation failed because the operation didn't complete within the allotted time.

UNKNOWN_ERROR: Indicates that the JoinDomain operation failed due to another type of error.

describeSMBSettingsResponse_domainName :: Lens' DescribeSMBSettingsResponse (Maybe Text)

The name of the domain that the gateway is joined to.

describeSMBSettingsResponse_sMBGuestPasswordSet :: Lens' DescribeSMBSettingsResponse (Maybe Bool)

This value is true if a password for the guest user smbguest is set, otherwise false. Only supported for S3 File Gateways.

Valid Values: true | false

describeSMBSettingsResponse_sMBSecurityStrategy :: Lens' DescribeSMBSettingsResponse (Maybe SMBSecurityStrategy)

The type of security strategy that was specified for file gateway.

ClientSpecified: If you use this option, requests are established based on what is negotiated by the client. This option is recommended when you want to maximize compatibility across different clients in your environment. Only supported for S3 File Gateways.

MandatorySigning: If you use this option, file gateway only allows connections from SMBv2 or SMBv3 clients that have signing enabled. This option works with SMB clients on Microsoft Windows Vista, Windows Server 2008 or newer.

MandatoryEncryption: If you use this option, file gateway only allows connections from SMBv3 clients that have encryption enabled. This option is highly recommended for environments that handle sensitive data. This option works with SMB clients on Microsoft Windows 8, Windows Server 2012 or newer.

describeSMBSettingsResponse_httpStatus :: Lens' DescribeSMBSettingsResponse Int

The response's http status code.