Network.AWS.StorageGateway.DescribeTapeRecoveryPoints

Description

Returns a list of virtual tape recovery points that are available for the specified tape gateway.

A recovery point is a point-in-time view of a virtual tape at which all the data on the virtual tape is consistent. If your gateway crashes, virtual tapes that have recovery points can be recovered to a new gateway. This operation is only supported in the tape gateway type.

This operation returns paginated results.

Creating a Request

data DescribeTapeRecoveryPoints

DescribeTapeRecoveryPointsInput

See: newDescribeTapeRecoveryPoints smart constructor.

Constructors

DescribeTapeRecoveryPoints'

Fields

marker :: Maybe Text

An opaque string that indicates the position at which to begin describing the virtual tape recovery points.

limit :: Maybe Natural

Specifies that the number of virtual tape recovery points that are described be limited to the specified number.

gatewayARN :: Text

newDescribeTapeRecoveryPoints

Arguments

:: Text
$sel:gatewayARN:DescribeTapeRecoveryPoints'

-> DescribeTapeRecoveryPoints

Create a value of DescribeTapeRecoveryPoints with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:DescribeTapeRecoveryPoints', describeTapeRecoveryPoints_marker - An opaque string that indicates the position at which to begin describing the virtual tape recovery points.

$sel:limit:DescribeTapeRecoveryPoints', describeTapeRecoveryPoints_limit - Specifies that the number of virtual tape recovery points that are described be limited to the specified number.

$sel:gatewayARN:DescribeTapeRecoveryPoints', describeTapeRecoveryPoints_gatewayARN - Undocumented member.

Request Lenses

describeTapeRecoveryPoints_marker :: Lens' DescribeTapeRecoveryPoints (Maybe Text)

An opaque string that indicates the position at which to begin describing the virtual tape recovery points.

describeTapeRecoveryPoints_limit :: Lens' DescribeTapeRecoveryPoints (Maybe Natural)

Specifies that the number of virtual tape recovery points that are described be limited to the specified number.

describeTapeRecoveryPoints_gatewayARN :: Lens' DescribeTapeRecoveryPoints Text

Undocumented member.

Destructuring the Response

data DescribeTapeRecoveryPointsResponse

DescribeTapeRecoveryPointsOutput

See: newDescribeTapeRecoveryPointsResponse smart constructor.

Constructors

DescribeTapeRecoveryPointsResponse'

Fields

tapeRecoveryPointInfos :: Maybe [TapeRecoveryPointInfo]

An array of TapeRecoveryPointInfos that are available for the specified gateway.

gatewayARN :: Maybe Text

marker :: Maybe Text

An opaque string that indicates the position at which the virtual tape recovery points that were listed for description ended.

Use this marker in your next request to list the next set of virtual tape recovery points in the list. If there are no more recovery points to describe, this field does not appear in the response.

httpStatus :: Int

The response's http status code.

newDescribeTapeRecoveryPointsResponse

Arguments

:: Int
$sel:httpStatus:DescribeTapeRecoveryPointsResponse'

-> DescribeTapeRecoveryPointsResponse

Create a value of DescribeTapeRecoveryPointsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tapeRecoveryPointInfos:DescribeTapeRecoveryPointsResponse', describeTapeRecoveryPointsResponse_tapeRecoveryPointInfos - An array of TapeRecoveryPointInfos that are available for the specified gateway.

$sel:gatewayARN:DescribeTapeRecoveryPoints', describeTapeRecoveryPointsResponse_gatewayARN - Undocumented member.

$sel:marker:DescribeTapeRecoveryPoints', describeTapeRecoveryPointsResponse_marker - An opaque string that indicates the position at which the virtual tape recovery points that were listed for description ended.

Use this marker in your next request to list the next set of virtual tape recovery points in the list. If there are no more recovery points to describe, this field does not appear in the response.

$sel:httpStatus:DescribeTapeRecoveryPointsResponse', describeTapeRecoveryPointsResponse_httpStatus - The response's http status code.

Response Lenses

describeTapeRecoveryPointsResponse_tapeRecoveryPointInfos :: Lens' DescribeTapeRecoveryPointsResponse (Maybe [TapeRecoveryPointInfo])

An array of TapeRecoveryPointInfos that are available for the specified gateway.

describeTapeRecoveryPointsResponse_gatewayARN :: Lens' DescribeTapeRecoveryPointsResponse (Maybe Text)

Undocumented member.

describeTapeRecoveryPointsResponse_marker :: Lens' DescribeTapeRecoveryPointsResponse (Maybe Text)

An opaque string that indicates the position at which the virtual tape recovery points that were listed for description ended.

Use this marker in your next request to list the next set of virtual tape recovery points in the list. If there are no more recovery points to describe, this field does not appear in the response.

describeTapeRecoveryPointsResponse_httpStatus :: Lens' DescribeTapeRecoveryPointsResponse Int

The response's http status code.