Network.AWS.StorageGateway.DescribeTapes

Description

Returns a description of the specified Amazon Resource Name (ARN) of virtual tapes. If a TapeARN is not specified, returns a description of all virtual tapes associated with the specified gateway. This operation is only supported in the tape gateway type.

This operation returns paginated results.

Creating a Request

data DescribeTapes

DescribeTapesInput

See: newDescribeTapes smart constructor.

Constructors

DescribeTapes'

Fields

marker :: Maybe Text

A marker value, obtained in a previous call to DescribeTapes. This marker indicates which page of results to retrieve.

If not specified, the first page of results is retrieved.

limit :: Maybe Natural

Specifies that the number of virtual tapes described be limited to the specified number.

Amazon Web Services may impose its own limit, if this field is not set.

tapeARNs :: Maybe [Text]

Specifies one or more unique Amazon Resource Names (ARNs) that represent the virtual tapes you want to describe. If this parameter is not specified, Tape gateway returns a description of all virtual tapes associated with the specified gateway.

gatewayARN :: Text

newDescribeTapes

Arguments

:: Text
$sel:gatewayARN:DescribeTapes'

-> DescribeTapes

Create a value of DescribeTapes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:DescribeTapes', describeTapes_marker - A marker value, obtained in a previous call to DescribeTapes. This marker indicates which page of results to retrieve.

If not specified, the first page of results is retrieved.

$sel:limit:DescribeTapes', describeTapes_limit - Specifies that the number of virtual tapes described be limited to the specified number.

Amazon Web Services may impose its own limit, if this field is not set.

$sel:tapeARNs:DescribeTapes', describeTapes_tapeARNs - Specifies one or more unique Amazon Resource Names (ARNs) that represent the virtual tapes you want to describe. If this parameter is not specified, Tape gateway returns a description of all virtual tapes associated with the specified gateway.

$sel:gatewayARN:DescribeTapes', describeTapes_gatewayARN - Undocumented member.

Request Lenses

describeTapes_marker :: Lens' DescribeTapes (Maybe Text)

A marker value, obtained in a previous call to DescribeTapes. This marker indicates which page of results to retrieve.

If not specified, the first page of results is retrieved.

describeTapes_limit :: Lens' DescribeTapes (Maybe Natural)

Specifies that the number of virtual tapes described be limited to the specified number.

Amazon Web Services may impose its own limit, if this field is not set.

describeTapes_tapeARNs :: Lens' DescribeTapes (Maybe [Text])

Specifies one or more unique Amazon Resource Names (ARNs) that represent the virtual tapes you want to describe. If this parameter is not specified, Tape gateway returns a description of all virtual tapes associated with the specified gateway.

describeTapes_gatewayARN :: Lens' DescribeTapes Text

Undocumented member.

Destructuring the Response

data DescribeTapesResponse

DescribeTapesOutput

See: newDescribeTapesResponse smart constructor.

Constructors

DescribeTapesResponse'

Fields

marker :: Maybe Text

An opaque string that can be used as part of a subsequent DescribeTapes call to retrieve the next page of results.

If a response does not contain a marker, then there are no more results to be retrieved.

tapes :: Maybe [Tape]

An array of virtual tape descriptions.

httpStatus :: Int

The response's http status code.

newDescribeTapesResponse

Arguments

:: Int
$sel:httpStatus:DescribeTapesResponse'

-> DescribeTapesResponse

Create a value of DescribeTapesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:DescribeTapes', describeTapesResponse_marker - An opaque string that can be used as part of a subsequent DescribeTapes call to retrieve the next page of results.

If a response does not contain a marker, then there are no more results to be retrieved.

$sel:tapes:DescribeTapesResponse', describeTapesResponse_tapes - An array of virtual tape descriptions.

$sel:httpStatus:DescribeTapesResponse', describeTapesResponse_httpStatus - The response's http status code.

Response Lenses

describeTapesResponse_marker :: Lens' DescribeTapesResponse (Maybe Text)

An opaque string that can be used as part of a subsequent DescribeTapes call to retrieve the next page of results.

If a response does not contain a marker, then there are no more results to be retrieved.

describeTapesResponse_tapes :: Lens' DescribeTapesResponse (Maybe [Tape])

An array of virtual tape descriptions.

describeTapesResponse_httpStatus :: Lens' DescribeTapesResponse Int

The response's http status code.