Network.AWS.StorageGateway.DetachVolume

Description

Disconnects a volume from an iSCSI connection and then detaches the volume from the specified gateway. Detaching and attaching a volume enables you to recover your data from one gateway to a different gateway without creating a snapshot. It also makes it easier to move your volumes from an on-premises gateway to a gateway hosted on an Amazon EC2 instance. This operation is only supported in the volume gateway type.

Creating a Request

data DetachVolume

AttachVolumeInput

See: newDetachVolume smart constructor.

Constructors

DetachVolume'

Fields

forceDetach :: Maybe Bool

Set to true to forcibly remove the iSCSI connection of the target volume and detach the volume. The default is false. If this value is set to false, you must manually disconnect the iSCSI connection from the target volume.

Valid Values: true | false

volumeARN :: Text

The Amazon Resource Name (ARN) of the volume to detach from the gateway.

newDetachVolume

Arguments

:: Text
$sel:volumeARN:DetachVolume'

-> DetachVolume

Create a value of DetachVolume with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:forceDetach:DetachVolume', detachVolume_forceDetach - Set to true to forcibly remove the iSCSI connection of the target volume and detach the volume. The default is false. If this value is set to false, you must manually disconnect the iSCSI connection from the target volume.

Valid Values: true | false

$sel:volumeARN:DetachVolume', detachVolume_volumeARN - The Amazon Resource Name (ARN) of the volume to detach from the gateway.

Request Lenses

detachVolume_forceDetach :: Lens' DetachVolume (Maybe Bool)

Set to true to forcibly remove the iSCSI connection of the target volume and detach the volume. The default is false. If this value is set to false, you must manually disconnect the iSCSI connection from the target volume.

Valid Values: true | false

detachVolume_volumeARN :: Lens' DetachVolume Text

The Amazon Resource Name (ARN) of the volume to detach from the gateway.

Destructuring the Response

data DetachVolumeResponse

AttachVolumeOutput

See: newDetachVolumeResponse smart constructor.

Constructors

DetachVolumeResponse'

Fields

volumeARN :: Maybe Text

The Amazon Resource Name (ARN) of the volume that was detached.

httpStatus :: Int

The response's http status code.

newDetachVolumeResponse

Arguments

:: Int
$sel:httpStatus:DetachVolumeResponse'

-> DetachVolumeResponse

Create a value of DetachVolumeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:volumeARN:DetachVolume', detachVolumeResponse_volumeARN - The Amazon Resource Name (ARN) of the volume that was detached.

$sel:httpStatus:DetachVolumeResponse', detachVolumeResponse_httpStatus - The response's http status code.

Response Lenses

detachVolumeResponse_volumeARN :: Lens' DetachVolumeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the volume that was detached.

detachVolumeResponse_httpStatus :: Lens' DetachVolumeResponse Int

The response's http status code.