Network.AWS.StorageGateway.ListFileShares

Description

Gets a list of the file shares for a specific S3 File Gateway, or the list of file shares that belong to the calling user account. This operation is only supported for S3 File Gateways.

This operation returns paginated results.

Creating a Request

data ListFileShares

ListFileShareInput

See: newListFileShares smart constructor.

Constructors

ListFileShares'

Fields

gatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway whose file shares you want to list. If this field is not present, all file shares under your account are listed.

marker :: Maybe Text

Opaque pagination token returned from a previous ListFileShares operation. If present, Marker specifies where to continue the list from after a previous call to ListFileShares. Optional.

limit :: Maybe Natural

The maximum number of file shares to return in the response. The value must be an integer with a value greater than zero. Optional.

newListFileShares :: ListFileShares

Create a value of ListFileShares with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:ListFileShares', listFileShares_gatewayARN - The Amazon Resource Name (ARN) of the gateway whose file shares you want to list. If this field is not present, all file shares under your account are listed.

$sel:marker:ListFileShares', listFileShares_marker - Opaque pagination token returned from a previous ListFileShares operation. If present, Marker specifies where to continue the list from after a previous call to ListFileShares. Optional.

$sel:limit:ListFileShares', listFileShares_limit - The maximum number of file shares to return in the response. The value must be an integer with a value greater than zero. Optional.

Request Lenses

listFileShares_gatewayARN :: Lens' ListFileShares (Maybe Text)

The Amazon Resource Name (ARN) of the gateway whose file shares you want to list. If this field is not present, all file shares under your account are listed.

listFileShares_marker :: Lens' ListFileShares (Maybe Text)

Opaque pagination token returned from a previous ListFileShares operation. If present, Marker specifies where to continue the list from after a previous call to ListFileShares. Optional.

listFileShares_limit :: Lens' ListFileShares (Maybe Natural)

The maximum number of file shares to return in the response. The value must be an integer with a value greater than zero. Optional.

Destructuring the Response

data ListFileSharesResponse

ListFileShareOutput

See: newListFileSharesResponse smart constructor.

Constructors

ListFileSharesResponse'

Fields

fileShareInfoList :: Maybe [FileShareInfo]

An array of information about the S3 File Gateway's file shares.

marker :: Maybe Text

If the request includes Marker, the response returns that value in this field.

nextMarker :: Maybe Text

If a value is present, there are more file shares to return. In a subsequent request, use NextMarker as the value for Marker to retrieve the next set of file shares.

httpStatus :: Int

The response's http status code.

newListFileSharesResponse

Arguments

:: Int
$sel:httpStatus:ListFileSharesResponse'

-> ListFileSharesResponse

Create a value of ListFileSharesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fileShareInfoList:ListFileSharesResponse', listFileSharesResponse_fileShareInfoList - An array of information about the S3 File Gateway's file shares.

$sel:marker:ListFileShares', listFileSharesResponse_marker - If the request includes Marker, the response returns that value in this field.

$sel:nextMarker:ListFileSharesResponse', listFileSharesResponse_nextMarker - If a value is present, there are more file shares to return. In a subsequent request, use NextMarker as the value for Marker to retrieve the next set of file shares.

$sel:httpStatus:ListFileSharesResponse', listFileSharesResponse_httpStatus - The response's http status code.

Response Lenses

listFileSharesResponse_fileShareInfoList :: Lens' ListFileSharesResponse (Maybe [FileShareInfo])

An array of information about the S3 File Gateway's file shares.

listFileSharesResponse_marker :: Lens' ListFileSharesResponse (Maybe Text)

If the request includes Marker, the response returns that value in this field.

listFileSharesResponse_nextMarker :: Lens' ListFileSharesResponse (Maybe Text)

If a value is present, there are more file shares to return. In a subsequent request, use NextMarker as the value for Marker to retrieve the next set of file shares.

listFileSharesResponse_httpStatus :: Lens' ListFileSharesResponse Int

The response's http status code.