Network.AWS.StorageGateway.ListFileSystemAssociations

Description

Gets a list of FileSystemAssociationSummary objects. Each object contains a summary of a file system association. This operation is only supported for FSx File Gateways.

This operation returns paginated results.

Creating a Request

data ListFileSystemAssociations

See: newListFileSystemAssociations smart constructor.

Constructors

ListFileSystemAssociations'

Fields

gatewayARN :: Maybe Text

marker :: Maybe Text

Opaque pagination token returned from a previous ListFileSystemAssociations operation. If present, Marker specifies where to continue the list from after a previous call to ListFileSystemAssociations. Optional.

limit :: Maybe Natural

The maximum number of file system associations to return in the response. If present, Limit must be an integer with a value greater than zero. Optional.

newListFileSystemAssociations :: ListFileSystemAssociations

Create a value of ListFileSystemAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:ListFileSystemAssociations', listFileSystemAssociations_gatewayARN - Undocumented member.

$sel:marker:ListFileSystemAssociations', listFileSystemAssociations_marker - Opaque pagination token returned from a previous ListFileSystemAssociations operation. If present, Marker specifies where to continue the list from after a previous call to ListFileSystemAssociations. Optional.

$sel:limit:ListFileSystemAssociations', listFileSystemAssociations_limit - The maximum number of file system associations to return in the response. If present, Limit must be an integer with a value greater than zero. Optional.

Request Lenses

listFileSystemAssociations_gatewayARN :: Lens' ListFileSystemAssociations (Maybe Text)

Undocumented member.

listFileSystemAssociations_marker :: Lens' ListFileSystemAssociations (Maybe Text)

Opaque pagination token returned from a previous ListFileSystemAssociations operation. If present, Marker specifies where to continue the list from after a previous call to ListFileSystemAssociations. Optional.

listFileSystemAssociations_limit :: Lens' ListFileSystemAssociations (Maybe Natural)

The maximum number of file system associations to return in the response. If present, Limit must be an integer with a value greater than zero. Optional.

Destructuring the Response

data ListFileSystemAssociationsResponse

See: newListFileSystemAssociationsResponse smart constructor.

Constructors

ListFileSystemAssociationsResponse'

Fields

fileSystemAssociationSummaryList :: Maybe [FileSystemAssociationSummary]

An array of information about the Amazon FSx gateway's file system associations.

marker :: Maybe Text

If the request includes Marker, the response returns that value in this field.

nextMarker :: Maybe Text

If a value is present, there are more file system associations to return. In a subsequent request, use NextMarker as the value for Marker to retrieve the next set of file system associations.

httpStatus :: Int

The response's http status code.

newListFileSystemAssociationsResponse

Arguments

:: Int
$sel:httpStatus:ListFileSystemAssociationsResponse'

-> ListFileSystemAssociationsResponse

Create a value of ListFileSystemAssociationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fileSystemAssociationSummaryList:ListFileSystemAssociationsResponse', listFileSystemAssociationsResponse_fileSystemAssociationSummaryList - An array of information about the Amazon FSx gateway's file system associations.

$sel:marker:ListFileSystemAssociations', listFileSystemAssociationsResponse_marker - If the request includes Marker, the response returns that value in this field.

$sel:nextMarker:ListFileSystemAssociationsResponse', listFileSystemAssociationsResponse_nextMarker - If a value is present, there are more file system associations to return. In a subsequent request, use NextMarker as the value for Marker to retrieve the next set of file system associations.

$sel:httpStatus:ListFileSystemAssociationsResponse', listFileSystemAssociationsResponse_httpStatus - The response's http status code.

Response Lenses

listFileSystemAssociationsResponse_fileSystemAssociationSummaryList :: Lens' ListFileSystemAssociationsResponse (Maybe [FileSystemAssociationSummary])

An array of information about the Amazon FSx gateway's file system associations.

listFileSystemAssociationsResponse_marker :: Lens' ListFileSystemAssociationsResponse (Maybe Text)

If the request includes Marker, the response returns that value in this field.

listFileSystemAssociationsResponse_nextMarker :: Lens' ListFileSystemAssociationsResponse (Maybe Text)

If a value is present, there are more file system associations to return. In a subsequent request, use NextMarker as the value for Marker to retrieve the next set of file system associations.

listFileSystemAssociationsResponse_httpStatus :: Lens' ListFileSystemAssociationsResponse Int

The response's http status code.