Network.AWS.StorageGateway.RefreshCache

Description

Refreshes the cached inventory of objects for the specified file share. This operation finds objects in the Amazon S3 bucket that were added, removed, or replaced since the gateway last listed the bucket's contents and cached the results. This operation does not import files into the S3 File Gateway cache storage. It only updates the cached inventory to reflect changes in the inventory of the objects in the S3 bucket. This operation is only supported in the S3 File Gateway types.

You can subscribe to be notified through an Amazon CloudWatch event when your RefreshCache operation completes. For more information, see Getting notified about file operations in the Storage Gateway User Guide. This operation is Only supported for S3 File Gateways.

When this API is called, it only initiates the refresh operation. When the API call completes and returns a success code, it doesn't necessarily mean that the file refresh has completed. You should use the refresh-complete notification to determine that the operation has completed before you check for new files on the gateway file share. You can subscribe to be notified through a CloudWatch event when your RefreshCache operation completes.

Throttle limit: This API is asynchronous, so the gateway will accept no more than two refreshes at any time. We recommend using the refresh-complete CloudWatch event notification before issuing additional requests. For more information, see Getting notified about file operations in the Storage Gateway User Guide.

If you invoke the RefreshCache API when two requests are already being processed, any new request will cause an InvalidGatewayRequestException error because too many requests were sent to the server.

For more information, see Getting notified about file operations in the Storage Gateway User Guide.

Creating a Request

data RefreshCache

RefreshCacheInput

See: newRefreshCache smart constructor.

Constructors

RefreshCache'

Fields

folderList :: Maybe (NonEmpty Text)

A comma-separated list of the paths of folders to refresh in the cache. The default is ["/"]. The default refreshes objects and folders at the root of the Amazon S3 bucket. If Recursive is set to true, the entire S3 bucket that the file share has access to is refreshed.

recursive :: Maybe Bool

A value that specifies whether to recursively refresh folders in the cache. The refresh includes folders that were in the cache the last time the gateway listed the folder's contents. If this value set to true, each folder that is listed in FolderList is recursively updated. Otherwise, subfolders listed in FolderList are not refreshed. Only objects that are in folders listed directly under FolderList are found and used for the update. The default is true.

Valid Values: true | false

fileShareARN :: Text

The Amazon Resource Name (ARN) of the file share you want to refresh.

newRefreshCache

Arguments

:: Text
$sel:fileShareARN:RefreshCache'

-> RefreshCache

Create a value of RefreshCache with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:folderList:RefreshCache', refreshCache_folderList - A comma-separated list of the paths of folders to refresh in the cache. The default is ["/"]. The default refreshes objects and folders at the root of the Amazon S3 bucket. If Recursive is set to true, the entire S3 bucket that the file share has access to is refreshed.

$sel:recursive:RefreshCache', refreshCache_recursive - A value that specifies whether to recursively refresh folders in the cache. The refresh includes folders that were in the cache the last time the gateway listed the folder's contents. If this value set to true, each folder that is listed in FolderList is recursively updated. Otherwise, subfolders listed in FolderList are not refreshed. Only objects that are in folders listed directly under FolderList are found and used for the update. The default is true.

Valid Values: true | false

$sel:fileShareARN:RefreshCache', refreshCache_fileShareARN - The Amazon Resource Name (ARN) of the file share you want to refresh.

Request Lenses

refreshCache_folderList :: Lens' RefreshCache (Maybe (NonEmpty Text))

A comma-separated list of the paths of folders to refresh in the cache. The default is ["/"]. The default refreshes objects and folders at the root of the Amazon S3 bucket. If Recursive is set to true, the entire S3 bucket that the file share has access to is refreshed.

refreshCache_recursive :: Lens' RefreshCache (Maybe Bool)

A value that specifies whether to recursively refresh folders in the cache. The refresh includes folders that were in the cache the last time the gateway listed the folder's contents. If this value set to true, each folder that is listed in FolderList is recursively updated. Otherwise, subfolders listed in FolderList are not refreshed. Only objects that are in folders listed directly under FolderList are found and used for the update. The default is true.

Valid Values: true | false

refreshCache_fileShareARN :: Lens' RefreshCache Text

The Amazon Resource Name (ARN) of the file share you want to refresh.

Destructuring the Response

data RefreshCacheResponse

RefreshCacheOutput

See: newRefreshCacheResponse smart constructor.

Constructors

RefreshCacheResponse'

Fields

fileShareARN :: Maybe Text

notificationId :: Maybe Text

httpStatus :: Int

The response's http status code.

newRefreshCacheResponse

Arguments

:: Int
$sel:httpStatus:RefreshCacheResponse'

-> RefreshCacheResponse

Create a value of RefreshCacheResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fileShareARN:RefreshCache', refreshCacheResponse_fileShareARN - Undocumented member.

$sel:notificationId:RefreshCacheResponse', refreshCacheResponse_notificationId - Undocumented member.

$sel:httpStatus:RefreshCacheResponse', refreshCacheResponse_httpStatus - The response's http status code.

Response Lenses

refreshCacheResponse_fileShareARN :: Lens' RefreshCacheResponse (Maybe Text)

Undocumented member.

refreshCacheResponse_notificationId :: Lens' RefreshCacheResponse (Maybe Text)

Undocumented member.

refreshCacheResponse_httpStatus :: Lens' RefreshCacheResponse Int

The response's http status code.