Network.AWS.StorageGateway.Types.BandwidthRateLimitInterval

Description

Documentation

data BandwidthRateLimitInterval

Describes a bandwidth rate limit interval for a gateway. A bandwidth rate limit schedule consists of one or more bandwidth rate limit intervals. A bandwidth rate limit interval defines a period of time on one or more days of the week, during which bandwidth rate limits are specified for uploading, downloading, or both.

See: newBandwidthRateLimitInterval smart constructor.

Constructors

BandwidthRateLimitInterval'

Fields

averageUploadRateLimitInBitsPerSec :: Maybe Natural

The average upload rate limit component of the bandwidth rate limit interval, in bits per second. This field does not appear in the response if the upload rate limit is not set.

averageDownloadRateLimitInBitsPerSec :: Maybe Natural

The average download rate limit component of the bandwidth rate limit interval, in bits per second. This field does not appear in the response if the download rate limit is not set.

startHourOfDay :: Natural

The hour of the day to start the bandwidth rate limit interval.

startMinuteOfHour :: Natural

The minute of the hour to start the bandwidth rate limit interval. The interval begins at the start of that minute. To begin an interval exactly at the start of the hour, use the value 0.

endHourOfDay :: Natural

The hour of the day to end the bandwidth rate limit interval.

endMinuteOfHour :: Natural

The minute of the hour to end the bandwidth rate limit interval.

The bandwidth rate limit interval ends at the end of the minute. To end an interval at the end of an hour, use the value 59.

daysOfWeek :: NonEmpty Natural

The days of the week component of the bandwidth rate limit interval, represented as ordinal numbers from 0 to 6, where 0 represents Sunday and 6 represents Saturday.

newBandwidthRateLimitInterval

Arguments

:: Natural
$sel:startHourOfDay:BandwidthRateLimitInterval'

-> Natural
$sel:startMinuteOfHour:BandwidthRateLimitInterval'

-> Natural
$sel:endHourOfDay:BandwidthRateLimitInterval'

-> Natural
$sel:endMinuteOfHour:BandwidthRateLimitInterval'

-> NonEmpty Natural
$sel:daysOfWeek:BandwidthRateLimitInterval'

-> BandwidthRateLimitInterval

Create a value of BandwidthRateLimitInterval with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:averageUploadRateLimitInBitsPerSec:BandwidthRateLimitInterval', bandwidthRateLimitInterval_averageUploadRateLimitInBitsPerSec - The average upload rate limit component of the bandwidth rate limit interval, in bits per second. This field does not appear in the response if the upload rate limit is not set.

$sel:averageDownloadRateLimitInBitsPerSec:BandwidthRateLimitInterval', bandwidthRateLimitInterval_averageDownloadRateLimitInBitsPerSec - The average download rate limit component of the bandwidth rate limit interval, in bits per second. This field does not appear in the response if the download rate limit is not set.

$sel:startHourOfDay:BandwidthRateLimitInterval', bandwidthRateLimitInterval_startHourOfDay - The hour of the day to start the bandwidth rate limit interval.

$sel:startMinuteOfHour:BandwidthRateLimitInterval', bandwidthRateLimitInterval_startMinuteOfHour - The minute of the hour to start the bandwidth rate limit interval. The interval begins at the start of that minute. To begin an interval exactly at the start of the hour, use the value 0.

$sel:endHourOfDay:BandwidthRateLimitInterval', bandwidthRateLimitInterval_endHourOfDay - The hour of the day to end the bandwidth rate limit interval.

$sel:endMinuteOfHour:BandwidthRateLimitInterval', bandwidthRateLimitInterval_endMinuteOfHour - The minute of the hour to end the bandwidth rate limit interval.

The bandwidth rate limit interval ends at the end of the minute. To end an interval at the end of an hour, use the value 59.

$sel:daysOfWeek:BandwidthRateLimitInterval', bandwidthRateLimitInterval_daysOfWeek - The days of the week component of the bandwidth rate limit interval, represented as ordinal numbers from 0 to 6, where 0 represents Sunday and 6 represents Saturday.

bandwidthRateLimitInterval_averageUploadRateLimitInBitsPerSec :: Lens' BandwidthRateLimitInterval (Maybe Natural)

The average upload rate limit component of the bandwidth rate limit interval, in bits per second. This field does not appear in the response if the upload rate limit is not set.

bandwidthRateLimitInterval_averageDownloadRateLimitInBitsPerSec :: Lens' BandwidthRateLimitInterval (Maybe Natural)

The average download rate limit component of the bandwidth rate limit interval, in bits per second. This field does not appear in the response if the download rate limit is not set.

bandwidthRateLimitInterval_startHourOfDay :: Lens' BandwidthRateLimitInterval Natural

The hour of the day to start the bandwidth rate limit interval.

bandwidthRateLimitInterval_startMinuteOfHour :: Lens' BandwidthRateLimitInterval Natural

The minute of the hour to start the bandwidth rate limit interval. The interval begins at the start of that minute. To begin an interval exactly at the start of the hour, use the value 0.

bandwidthRateLimitInterval_endHourOfDay :: Lens' BandwidthRateLimitInterval Natural

The hour of the day to end the bandwidth rate limit interval.

bandwidthRateLimitInterval_endMinuteOfHour :: Lens' BandwidthRateLimitInterval Natural

The minute of the hour to end the bandwidth rate limit interval.

The bandwidth rate limit interval ends at the end of the minute. To end an interval at the end of an hour, use the value 59.

bandwidthRateLimitInterval_daysOfWeek :: Lens' BandwidthRateLimitInterval (NonEmpty Natural)

The days of the week component of the bandwidth rate limit interval, represented as ordinal numbers from 0 to 6, where 0 represents Sunday and 6 represents Saturday.