Network.AWS.StorageGateway.Types.CachediSCSIVolume

Description

Documentation

data CachediSCSIVolume

Describes an iSCSI cached volume.

See: newCachediSCSIVolume smart constructor.

Constructors

CachediSCSIVolume'

Fields

volumeiSCSIAttributes :: Maybe VolumeiSCSIAttributes

An VolumeiSCSIAttributes object that represents a collection of iSCSI attributes for one stored volume.

volumeStatus :: Maybe Text

One of the VolumeStatus values that indicates the state of the storage volume.

sourceSnapshotId :: Maybe Text

If the cached volume was created from a snapshot, this field contains the snapshot ID used, e.g., snap-78e22663. Otherwise, this field is not included.

kmsKey :: Maybe Text

volumeAttachmentStatus :: Maybe Text

A value that indicates whether a storage volume is attached to or detached from a gateway. For more information, see Moving your volumes to a different gateway.

volumeARN :: Maybe Text

The Amazon Resource Name (ARN) of the storage volume.

volumeProgress :: Maybe Double

Represents the percentage complete if the volume is restoring or bootstrapping that represents the percent of data transferred. This field does not appear in the response if the cached volume is not restoring or bootstrapping.

volumeSizeInBytes :: Maybe Integer

The size, in bytes, of the volume capacity.

volumeUsedInBytes :: Maybe Integer

The size of the data stored on the volume in bytes. This value is calculated based on the number of blocks that are touched, instead of the actual amount of data written. This value can be useful for sequential write patterns but less accurate for random write patterns. VolumeUsedInBytes is different from the compressed size of the volume, which is the value that is used to calculate your bill.

This value is not available for volumes created prior to May 13, 2015, until you store data on the volume.

createdDate :: Maybe POSIX

The date the volume was created. Volumes created prior to March 28, 2017 don’t have this timestamp.

volumeId :: Maybe Text

The unique identifier of the volume, e.g., vol-AE4B946D.

volumeType :: Maybe Text

One of the VolumeType enumeration values that describes the type of the volume.

targetName :: Maybe Text

The name of the iSCSI target used by an initiator to connect to a volume and used as a suffix for the target ARN. For example, specifying TargetName as myvolume results in the target ARN of arn:aws:storagegateway:us-east-2:111122223333:gateway/sgw-12A3456B/target/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes on a gateway.

If you don't specify a value, Storage Gateway uses the value that was previously used for this volume as the new target name.

newCachediSCSIVolume :: CachediSCSIVolume

Create a value of CachediSCSIVolume with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:volumeiSCSIAttributes:CachediSCSIVolume', cachediSCSIVolume_volumeiSCSIAttributes - An VolumeiSCSIAttributes object that represents a collection of iSCSI attributes for one stored volume.

$sel:volumeStatus:CachediSCSIVolume', cachediSCSIVolume_volumeStatus - One of the VolumeStatus values that indicates the state of the storage volume.

$sel:sourceSnapshotId:CachediSCSIVolume', cachediSCSIVolume_sourceSnapshotId - If the cached volume was created from a snapshot, this field contains the snapshot ID used, e.g., snap-78e22663. Otherwise, this field is not included.

$sel:kmsKey:CachediSCSIVolume', cachediSCSIVolume_kmsKey - Undocumented member.

$sel:volumeAttachmentStatus:CachediSCSIVolume', cachediSCSIVolume_volumeAttachmentStatus - A value that indicates whether a storage volume is attached to or detached from a gateway. For more information, see Moving your volumes to a different gateway.

$sel:volumeARN:CachediSCSIVolume', cachediSCSIVolume_volumeARN - The Amazon Resource Name (ARN) of the storage volume.

$sel:volumeProgress:CachediSCSIVolume', cachediSCSIVolume_volumeProgress - Represents the percentage complete if the volume is restoring or bootstrapping that represents the percent of data transferred. This field does not appear in the response if the cached volume is not restoring or bootstrapping.

$sel:volumeSizeInBytes:CachediSCSIVolume', cachediSCSIVolume_volumeSizeInBytes - The size, in bytes, of the volume capacity.

$sel:volumeUsedInBytes:CachediSCSIVolume', cachediSCSIVolume_volumeUsedInBytes - The size of the data stored on the volume in bytes. This value is calculated based on the number of blocks that are touched, instead of the actual amount of data written. This value can be useful for sequential write patterns but less accurate for random write patterns. VolumeUsedInBytes is different from the compressed size of the volume, which is the value that is used to calculate your bill.

This value is not available for volumes created prior to May 13, 2015, until you store data on the volume.

$sel:createdDate:CachediSCSIVolume', cachediSCSIVolume_createdDate - The date the volume was created. Volumes created prior to March 28, 2017 don’t have this timestamp.

$sel:volumeId:CachediSCSIVolume', cachediSCSIVolume_volumeId - The unique identifier of the volume, e.g., vol-AE4B946D.

$sel:volumeType:CachediSCSIVolume', cachediSCSIVolume_volumeType - One of the VolumeType enumeration values that describes the type of the volume.

$sel:targetName:CachediSCSIVolume', cachediSCSIVolume_targetName - The name of the iSCSI target used by an initiator to connect to a volume and used as a suffix for the target ARN. For example, specifying TargetName as myvolume results in the target ARN of arn:aws:storagegateway:us-east-2:111122223333:gateway/sgw-12A3456B/target/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes on a gateway.

If you don't specify a value, Storage Gateway uses the value that was previously used for this volume as the new target name.

cachediSCSIVolume_volumeiSCSIAttributes :: Lens' CachediSCSIVolume (Maybe VolumeiSCSIAttributes)

An VolumeiSCSIAttributes object that represents a collection of iSCSI attributes for one stored volume.

cachediSCSIVolume_volumeStatus :: Lens' CachediSCSIVolume (Maybe Text)

One of the VolumeStatus values that indicates the state of the storage volume.

cachediSCSIVolume_sourceSnapshotId :: Lens' CachediSCSIVolume (Maybe Text)

If the cached volume was created from a snapshot, this field contains the snapshot ID used, e.g., snap-78e22663. Otherwise, this field is not included.

cachediSCSIVolume_kmsKey :: Lens' CachediSCSIVolume (Maybe Text)

Undocumented member.

cachediSCSIVolume_volumeAttachmentStatus :: Lens' CachediSCSIVolume (Maybe Text)

A value that indicates whether a storage volume is attached to or detached from a gateway. For more information, see Moving your volumes to a different gateway.

cachediSCSIVolume_volumeARN :: Lens' CachediSCSIVolume (Maybe Text)

The Amazon Resource Name (ARN) of the storage volume.

cachediSCSIVolume_volumeProgress :: Lens' CachediSCSIVolume (Maybe Double)

Represents the percentage complete if the volume is restoring or bootstrapping that represents the percent of data transferred. This field does not appear in the response if the cached volume is not restoring or bootstrapping.

cachediSCSIVolume_volumeSizeInBytes :: Lens' CachediSCSIVolume (Maybe Integer)

The size, in bytes, of the volume capacity.

cachediSCSIVolume_volumeUsedInBytes :: Lens' CachediSCSIVolume (Maybe Integer)

The size of the data stored on the volume in bytes. This value is calculated based on the number of blocks that are touched, instead of the actual amount of data written. This value can be useful for sequential write patterns but less accurate for random write patterns. VolumeUsedInBytes is different from the compressed size of the volume, which is the value that is used to calculate your bill.

This value is not available for volumes created prior to May 13, 2015, until you store data on the volume.

cachediSCSIVolume_createdDate :: Lens' CachediSCSIVolume (Maybe UTCTime)

The date the volume was created. Volumes created prior to March 28, 2017 don’t have this timestamp.

cachediSCSIVolume_volumeId :: Lens' CachediSCSIVolume (Maybe Text)

The unique identifier of the volume, e.g., vol-AE4B946D.

cachediSCSIVolume_volumeType :: Lens' CachediSCSIVolume (Maybe Text)

One of the VolumeType enumeration values that describes the type of the volume.

cachediSCSIVolume_targetName :: Lens' CachediSCSIVolume (Maybe Text)

The name of the iSCSI target used by an initiator to connect to a volume and used as a suffix for the target ARN. For example, specifying TargetName as myvolume results in the target ARN of arn:aws:storagegateway:us-east-2:111122223333:gateway/sgw-12A3456B/target/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes on a gateway.

If you don't specify a value, Storage Gateway uses the value that was previously used for this volume as the new target name.