Network.AWS.StorageGateway.Types.ChapInfo

Description

Documentation

data ChapInfo

Describes Challenge-Handshake Authentication Protocol (CHAP) information that supports authentication between your gateway and iSCSI initiators.

See: newChapInfo smart constructor.

Constructors

ChapInfo'

Fields

targetARN :: Maybe Text

The Amazon Resource Name (ARN) of the volume.

Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

secretToAuthenticateInitiator :: Maybe (Sensitive Text)

The secret key that the initiator (for example, the Windows client) must provide to participate in mutual CHAP with the target.

initiatorName :: Maybe Text

The iSCSI initiator that connects to the target.

secretToAuthenticateTarget :: Maybe (Sensitive Text)

The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g., Windows client).

newChapInfo :: ChapInfo

Create a value of ChapInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetARN:ChapInfo', chapInfo_targetARN - The Amazon Resource Name (ARN) of the volume.

Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

$sel:secretToAuthenticateInitiator:ChapInfo', chapInfo_secretToAuthenticateInitiator - The secret key that the initiator (for example, the Windows client) must provide to participate in mutual CHAP with the target.

$sel:initiatorName:ChapInfo', chapInfo_initiatorName - The iSCSI initiator that connects to the target.

$sel:secretToAuthenticateTarget:ChapInfo', chapInfo_secretToAuthenticateTarget - The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g., Windows client).

chapInfo_targetARN :: Lens' ChapInfo (Maybe Text)

The Amazon Resource Name (ARN) of the volume.

Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

chapInfo_secretToAuthenticateInitiator :: Lens' ChapInfo (Maybe Text)

The secret key that the initiator (for example, the Windows client) must provide to participate in mutual CHAP with the target.

chapInfo_initiatorName :: Lens' ChapInfo (Maybe Text)

The iSCSI initiator that connects to the target.

chapInfo_secretToAuthenticateTarget :: Lens' ChapInfo (Maybe Text)

The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g., Windows client).