Network.AWS.StorageGateway.Types.FileSystemAssociationInfo

Description

Documentation

data FileSystemAssociationInfo

Describes the object returned by DescribeFileSystemAssociations that describes a created file system association.

See: newFileSystemAssociationInfo smart constructor.

Constructors

FileSystemAssociationInfo'

Fields

auditDestinationARN :: Maybe Text

The Amazon Resource Name (ARN) of the storage used for the audit logs.

fileSystemAssociationARN :: Maybe Text

The Amazon Resource Name (ARN) of the file system association.

gatewayARN :: Maybe Text

cacheAttributes :: Maybe CacheAttributes

endpointNetworkConfiguration :: Maybe EndpointNetworkConfiguration

Specifies network configuration information for the gateway associated with the Amazon FSx file system.

If multiple file systems are associated with this gateway, this parameter's IpAddresses field is required.

locationARN :: Maybe Text

The ARN of the backend Amazon FSx file system used for storing file data. For information, see FileSystem in the Amazon FSx API Reference.

fileSystemAssociationStatus :: Maybe Text

The status of the file system association. Valid Values: AVAILABLE | CREATING | DELETING | FORCE_DELETING | UPDATING | ERROR

tags :: Maybe [Tag]

A list of up to 50 tags assigned to the SMB file share, sorted alphabetically by key name. Each tag is a key-value pair.

newFileSystemAssociationInfo :: FileSystemAssociationInfo

Create a value of FileSystemAssociationInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:auditDestinationARN:FileSystemAssociationInfo', fileSystemAssociationInfo_auditDestinationARN - The Amazon Resource Name (ARN) of the storage used for the audit logs.

$sel:fileSystemAssociationARN:FileSystemAssociationInfo', fileSystemAssociationInfo_fileSystemAssociationARN - The Amazon Resource Name (ARN) of the file system association.

$sel:gatewayARN:FileSystemAssociationInfo', fileSystemAssociationInfo_gatewayARN - Undocumented member.

$sel:cacheAttributes:FileSystemAssociationInfo', fileSystemAssociationInfo_cacheAttributes - Undocumented member.

$sel:endpointNetworkConfiguration:FileSystemAssociationInfo', fileSystemAssociationInfo_endpointNetworkConfiguration - Specifies network configuration information for the gateway associated with the Amazon FSx file system.

If multiple file systems are associated with this gateway, this parameter's IpAddresses field is required.

$sel:locationARN:FileSystemAssociationInfo', fileSystemAssociationInfo_locationARN - The ARN of the backend Amazon FSx file system used for storing file data. For information, see FileSystem in the Amazon FSx API Reference.

$sel:fileSystemAssociationStatus:FileSystemAssociationInfo', fileSystemAssociationInfo_fileSystemAssociationStatus - The status of the file system association. Valid Values: AVAILABLE | CREATING | DELETING | FORCE_DELETING | UPDATING | ERROR

$sel:tags:FileSystemAssociationInfo', fileSystemAssociationInfo_tags - A list of up to 50 tags assigned to the SMB file share, sorted alphabetically by key name. Each tag is a key-value pair.

fileSystemAssociationInfo_auditDestinationARN :: Lens' FileSystemAssociationInfo (Maybe Text)

The Amazon Resource Name (ARN) of the storage used for the audit logs.

fileSystemAssociationInfo_fileSystemAssociationARN :: Lens' FileSystemAssociationInfo (Maybe Text)

The Amazon Resource Name (ARN) of the file system association.

fileSystemAssociationInfo_gatewayARN :: Lens' FileSystemAssociationInfo (Maybe Text)

Undocumented member.

fileSystemAssociationInfo_cacheAttributes :: Lens' FileSystemAssociationInfo (Maybe CacheAttributes)

Undocumented member.

fileSystemAssociationInfo_endpointNetworkConfiguration :: Lens' FileSystemAssociationInfo (Maybe EndpointNetworkConfiguration)

Specifies network configuration information for the gateway associated with the Amazon FSx file system.

If multiple file systems are associated with this gateway, this parameter's IpAddresses field is required.

fileSystemAssociationInfo_locationARN :: Lens' FileSystemAssociationInfo (Maybe Text)

The ARN of the backend Amazon FSx file system used for storing file data. For information, see FileSystem in the Amazon FSx API Reference.

fileSystemAssociationInfo_fileSystemAssociationStatus :: Lens' FileSystemAssociationInfo (Maybe Text)

The status of the file system association. Valid Values: AVAILABLE | CREATING | DELETING | FORCE_DELETING | UPDATING | ERROR

fileSystemAssociationInfo_tags :: Lens' FileSystemAssociationInfo (Maybe [Tag])

A list of up to 50 tags assigned to the SMB file share, sorted alphabetically by key name. Each tag is a key-value pair.