Network.AWS.StorageGateway.Types.NFSFileShareInfo

Description

Documentation

data NFSFileShareInfo

The Unix file permissions and ownership information assigned, by default, to native S3 objects when an S3 File Gateway discovers them in S3 buckets. This operation is only supported in S3 File Gateways.

See: newNFSFileShareInfo smart constructor.

Constructors

NFSFileShareInfo'

Fields

auditDestinationARN :: Maybe Text

The Amazon Resource Name (ARN) of the storage used for audit logs.

fileShareStatus :: Maybe Text

kmsKey :: Maybe Text

gatewayARN :: Maybe Text

path :: Maybe Text

vPCEndpointDNSName :: Maybe Text

Specifies the DNS name for the VPC endpoint that the NFS file share uses to connect to Amazon S3.

This parameter is required for NFS file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

cacheAttributes :: Maybe CacheAttributes

Refresh cache information for the file share.

objectACL :: Maybe ObjectACL

kmsEncrypted :: Maybe Bool

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

fileShareId :: Maybe Text

fileShareARN :: Maybe Text

defaultStorageClass :: Maybe Text

The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

fileShareName :: Maybe Text

The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN, or if an access point or access point alias is used.

role' :: Maybe Text

notificationPolicy :: Maybe Text

The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

squash :: Maybe Text

requesterPays :: Maybe Bool

A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

nFSFileShareDefaults :: Maybe NFSFileShareDefaults

locationARN :: Maybe Text

clientList :: Maybe (NonEmpty Text)

guessMIMETypeEnabled :: Maybe Bool

A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

readOnly :: Maybe Bool

A value that sets the write status of a file share. Set this value to true to set the write status to read-only, otherwise set to false.

Valid Values: true | false

bucketRegion :: Maybe Text

Specifies the Region of the S3 bucket where the NFS file share stores files.

This parameter is required for NFS file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

tags :: Maybe [Tag]

A list of up to 50 tags assigned to the NFS file share, sorted alphabetically by key name. Each tag is a key-value pair. For a gateway with more than 10 tags assigned, you can view all tags using the ListTagsForResource API operation.

newNFSFileShareInfo :: NFSFileShareInfo

Create a value of NFSFileShareInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:auditDestinationARN:NFSFileShareInfo', nFSFileShareInfo_auditDestinationARN - The Amazon Resource Name (ARN) of the storage used for audit logs.

$sel:fileShareStatus:NFSFileShareInfo', nFSFileShareInfo_fileShareStatus - Undocumented member.

$sel:kmsKey:NFSFileShareInfo', nFSFileShareInfo_kmsKey - Undocumented member.

$sel:gatewayARN:NFSFileShareInfo', nFSFileShareInfo_gatewayARN - Undocumented member.

$sel:path:NFSFileShareInfo', nFSFileShareInfo_path - Undocumented member.

$sel:vPCEndpointDNSName:NFSFileShareInfo', nFSFileShareInfo_vPCEndpointDNSName - Specifies the DNS name for the VPC endpoint that the NFS file share uses to connect to Amazon S3.

This parameter is required for NFS file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

$sel:cacheAttributes:NFSFileShareInfo', nFSFileShareInfo_cacheAttributes - Refresh cache information for the file share.

$sel:objectACL:NFSFileShareInfo', nFSFileShareInfo_objectACL - Undocumented member.

$sel:kmsEncrypted:NFSFileShareInfo', nFSFileShareInfo_kmsEncrypted - Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

$sel:fileShareId:NFSFileShareInfo', nFSFileShareInfo_fileShareId - Undocumented member.

$sel:fileShareARN:NFSFileShareInfo', nFSFileShareInfo_fileShareARN - Undocumented member.

$sel:defaultStorageClass:NFSFileShareInfo', nFSFileShareInfo_defaultStorageClass - The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

$sel:fileShareName:NFSFileShareInfo', nFSFileShareInfo_fileShareName - The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN, or if an access point or access point alias is used.

$sel:role':NFSFileShareInfo', nFSFileShareInfo_role - Undocumented member.

$sel:notificationPolicy:NFSFileShareInfo', nFSFileShareInfo_notificationPolicy - The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

$sel:squash:NFSFileShareInfo', nFSFileShareInfo_squash - Undocumented member.

$sel:requesterPays:NFSFileShareInfo', nFSFileShareInfo_requesterPays - A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

$sel:nFSFileShareDefaults:NFSFileShareInfo', nFSFileShareInfo_nFSFileShareDefaults - Undocumented member.

$sel:locationARN:NFSFileShareInfo', nFSFileShareInfo_locationARN - Undocumented member.

$sel:clientList:NFSFileShareInfo', nFSFileShareInfo_clientList - Undocumented member.

$sel:guessMIMETypeEnabled:NFSFileShareInfo', nFSFileShareInfo_guessMIMETypeEnabled - A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

$sel:readOnly:NFSFileShareInfo', nFSFileShareInfo_readOnly - A value that sets the write status of a file share. Set this value to true to set the write status to read-only, otherwise set to false.

Valid Values: true | false

$sel:bucketRegion:NFSFileShareInfo', nFSFileShareInfo_bucketRegion - Specifies the Region of the S3 bucket where the NFS file share stores files.

This parameter is required for NFS file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

$sel:tags:NFSFileShareInfo', nFSFileShareInfo_tags - A list of up to 50 tags assigned to the NFS file share, sorted alphabetically by key name. Each tag is a key-value pair. For a gateway with more than 10 tags assigned, you can view all tags using the ListTagsForResource API operation.

nFSFileShareInfo_auditDestinationARN :: Lens' NFSFileShareInfo (Maybe Text)

The Amazon Resource Name (ARN) of the storage used for audit logs.

nFSFileShareInfo_fileShareStatus :: Lens' NFSFileShareInfo (Maybe Text)

Undocumented member.

nFSFileShareInfo_kmsKey :: Lens' NFSFileShareInfo (Maybe Text)

Undocumented member.

nFSFileShareInfo_gatewayARN :: Lens' NFSFileShareInfo (Maybe Text)

Undocumented member.

nFSFileShareInfo_path :: Lens' NFSFileShareInfo (Maybe Text)

Undocumented member.

nFSFileShareInfo_vPCEndpointDNSName :: Lens' NFSFileShareInfo (Maybe Text)

Specifies the DNS name for the VPC endpoint that the NFS file share uses to connect to Amazon S3.

This parameter is required for NFS file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

nFSFileShareInfo_cacheAttributes :: Lens' NFSFileShareInfo (Maybe CacheAttributes)

Refresh cache information for the file share.

nFSFileShareInfo_objectACL :: Lens' NFSFileShareInfo (Maybe ObjectACL)

Undocumented member.

nFSFileShareInfo_kmsEncrypted :: Lens' NFSFileShareInfo (Maybe Bool)

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

nFSFileShareInfo_fileShareId :: Lens' NFSFileShareInfo (Maybe Text)

Undocumented member.

nFSFileShareInfo_fileShareARN :: Lens' NFSFileShareInfo (Maybe Text)

Undocumented member.

nFSFileShareInfo_defaultStorageClass :: Lens' NFSFileShareInfo (Maybe Text)

The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

nFSFileShareInfo_fileShareName :: Lens' NFSFileShareInfo (Maybe Text)

The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN, or if an access point or access point alias is used.

nFSFileShareInfo_role :: Lens' NFSFileShareInfo (Maybe Text)

Undocumented member.

nFSFileShareInfo_notificationPolicy :: Lens' NFSFileShareInfo (Maybe Text)

The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

nFSFileShareInfo_squash :: Lens' NFSFileShareInfo (Maybe Text)

Undocumented member.

nFSFileShareInfo_requesterPays :: Lens' NFSFileShareInfo (Maybe Bool)

A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

nFSFileShareInfo_nFSFileShareDefaults :: Lens' NFSFileShareInfo (Maybe NFSFileShareDefaults)

Undocumented member.

nFSFileShareInfo_locationARN :: Lens' NFSFileShareInfo (Maybe Text)

Undocumented member.

nFSFileShareInfo_clientList :: Lens' NFSFileShareInfo (Maybe (NonEmpty Text))

Undocumented member.

nFSFileShareInfo_guessMIMETypeEnabled :: Lens' NFSFileShareInfo (Maybe Bool)

A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

nFSFileShareInfo_readOnly :: Lens' NFSFileShareInfo (Maybe Bool)

A value that sets the write status of a file share. Set this value to true to set the write status to read-only, otherwise set to false.

Valid Values: true | false

nFSFileShareInfo_bucketRegion :: Lens' NFSFileShareInfo (Maybe Text)

Specifies the Region of the S3 bucket where the NFS file share stores files.

This parameter is required for NFS file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

nFSFileShareInfo_tags :: Lens' NFSFileShareInfo (Maybe [Tag])

A list of up to 50 tags assigned to the NFS file share, sorted alphabetically by key name. Each tag is a key-value pair. For a gateway with more than 10 tags assigned, you can view all tags using the ListTagsForResource API operation.