Network.AWS.StorageGateway.Types.SMBFileShareInfo

Description

Documentation

data SMBFileShareInfo

The Windows file permissions and ownership information assigned, by default, to native S3 objects when S3 File Gateway discovers them in S3 buckets. This operation is only supported for S3 File Gateways.

See: newSMBFileShareInfo smart constructor.

Constructors

SMBFileShareInfo'

Fields

accessBasedEnumeration :: Maybe Bool

Indicates whether AccessBasedEnumeration is enabled.

adminUserList :: Maybe [Text]

A list of users or groups in the Active Directory that have administrator rights to the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

auditDestinationARN :: Maybe Text

The Amazon Resource Name (ARN) of the storage used for audit logs.

fileShareStatus :: Maybe Text

invalidUserList :: Maybe [Text]

A list of users or groups in the Active Directory that are not allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

kmsKey :: Maybe Text

validUserList :: Maybe [Text]

A list of users or groups in the Active Directory that are allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

gatewayARN :: Maybe Text

path :: Maybe Text

The file share path used by the SMB client to identify the mount point.

vPCEndpointDNSName :: Maybe Text

Specifies the DNS name for the VPC endpoint that the SMB file share uses to connect to Amazon S3.

This parameter is required for SMB file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

authentication :: Maybe Text

cacheAttributes :: Maybe CacheAttributes

Refresh cache information for the file share.

objectACL :: Maybe ObjectACL

kmsEncrypted :: Maybe Bool

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

fileShareId :: Maybe Text

fileShareARN :: Maybe Text

defaultStorageClass :: Maybe Text

The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

fileShareName :: Maybe Text

The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN.

role' :: Maybe Text

sMBACLEnabled :: Maybe Bool

If this value is set to true, it indicates that access control list (ACL) is enabled on the SMB file share. If it is set to false, it indicates that file and directory permissions are mapped to the POSIX permission.

For more information, see Using Microsoft Windows ACLs to control access to an SMB file share in the Storage Gateway User Guide.

oplocksEnabled :: Maybe Bool

Specifies whether opportunistic locking is enabled for the SMB file share.

Enabling opportunistic locking on case-sensitive shares is not recommended for workloads that involve access to files with the same name in different case.

Valid Values: true | false

notificationPolicy :: Maybe Text

The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

requesterPays :: Maybe Bool

A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

locationARN :: Maybe Text

guessMIMETypeEnabled :: Maybe Bool

A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

readOnly :: Maybe Bool

A value that sets the write status of a file share. Set this value to true to set the write status to read-only, otherwise set to false.

Valid Values: true | false

bucketRegion :: Maybe Text

Specifies the Region of the S3 bucket where the SMB file share stores files.

This parameter is required for SMB file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

caseSensitivity :: Maybe CaseSensitivity

The case of an object name in an Amazon S3 bucket. For ClientSpecified, the client determines the case sensitivity. For CaseSensitive, the gateway determines the case sensitivity. The default value is ClientSpecified.

tags :: Maybe [Tag]

A list of up to 50 tags assigned to the SMB file share, sorted alphabetically by key name. Each tag is a key-value pair. For a gateway with more than 10 tags assigned, you can view all tags using the ListTagsForResource API operation.

newSMBFileShareInfo :: SMBFileShareInfo

Create a value of SMBFileShareInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessBasedEnumeration:SMBFileShareInfo', sMBFileShareInfo_accessBasedEnumeration - Indicates whether AccessBasedEnumeration is enabled.

$sel:adminUserList:SMBFileShareInfo', sMBFileShareInfo_adminUserList - A list of users or groups in the Active Directory that have administrator rights to the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

$sel:auditDestinationARN:SMBFileShareInfo', sMBFileShareInfo_auditDestinationARN - The Amazon Resource Name (ARN) of the storage used for audit logs.

$sel:fileShareStatus:SMBFileShareInfo', sMBFileShareInfo_fileShareStatus - Undocumented member.

$sel:invalidUserList:SMBFileShareInfo', sMBFileShareInfo_invalidUserList - A list of users or groups in the Active Directory that are not allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

$sel:kmsKey:SMBFileShareInfo', sMBFileShareInfo_kmsKey - Undocumented member.

$sel:validUserList:SMBFileShareInfo', sMBFileShareInfo_validUserList - A list of users or groups in the Active Directory that are allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

$sel:gatewayARN:SMBFileShareInfo', sMBFileShareInfo_gatewayARN - Undocumented member.

$sel:path:SMBFileShareInfo', sMBFileShareInfo_path - The file share path used by the SMB client to identify the mount point.

$sel:vPCEndpointDNSName:SMBFileShareInfo', sMBFileShareInfo_vPCEndpointDNSName - Specifies the DNS name for the VPC endpoint that the SMB file share uses to connect to Amazon S3.

This parameter is required for SMB file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

$sel:authentication:SMBFileShareInfo', sMBFileShareInfo_authentication - Undocumented member.

$sel:cacheAttributes:SMBFileShareInfo', sMBFileShareInfo_cacheAttributes - Refresh cache information for the file share.

$sel:objectACL:SMBFileShareInfo', sMBFileShareInfo_objectACL - Undocumented member.

$sel:kmsEncrypted:SMBFileShareInfo', sMBFileShareInfo_kmsEncrypted - Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

$sel:fileShareId:SMBFileShareInfo', sMBFileShareInfo_fileShareId - Undocumented member.

$sel:fileShareARN:SMBFileShareInfo', sMBFileShareInfo_fileShareARN - Undocumented member.

$sel:defaultStorageClass:SMBFileShareInfo', sMBFileShareInfo_defaultStorageClass - The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

$sel:fileShareName:SMBFileShareInfo', sMBFileShareInfo_fileShareName - The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN.

$sel:role':SMBFileShareInfo', sMBFileShareInfo_role - Undocumented member.

$sel:sMBACLEnabled:SMBFileShareInfo', sMBFileShareInfo_sMBACLEnabled - If this value is set to true, it indicates that access control list (ACL) is enabled on the SMB file share. If it is set to false, it indicates that file and directory permissions are mapped to the POSIX permission.

For more information, see Using Microsoft Windows ACLs to control access to an SMB file share in the Storage Gateway User Guide.

$sel:oplocksEnabled:SMBFileShareInfo', sMBFileShareInfo_oplocksEnabled - Specifies whether opportunistic locking is enabled for the SMB file share.

Enabling opportunistic locking on case-sensitive shares is not recommended for workloads that involve access to files with the same name in different case.

Valid Values: true | false

$sel:notificationPolicy:SMBFileShareInfo', sMBFileShareInfo_notificationPolicy - The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

$sel:requesterPays:SMBFileShareInfo', sMBFileShareInfo_requesterPays - A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

$sel:locationARN:SMBFileShareInfo', sMBFileShareInfo_locationARN - Undocumented member.

$sel:guessMIMETypeEnabled:SMBFileShareInfo', sMBFileShareInfo_guessMIMETypeEnabled - A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

$sel:readOnly:SMBFileShareInfo', sMBFileShareInfo_readOnly - A value that sets the write status of a file share. Set this value to true to set the write status to read-only, otherwise set to false.

Valid Values: true | false

$sel:bucketRegion:SMBFileShareInfo', sMBFileShareInfo_bucketRegion - Specifies the Region of the S3 bucket where the SMB file share stores files.

This parameter is required for SMB file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

$sel:caseSensitivity:SMBFileShareInfo', sMBFileShareInfo_caseSensitivity - The case of an object name in an Amazon S3 bucket. For ClientSpecified, the client determines the case sensitivity. For CaseSensitive, the gateway determines the case sensitivity. The default value is ClientSpecified.

$sel:tags:SMBFileShareInfo', sMBFileShareInfo_tags - A list of up to 50 tags assigned to the SMB file share, sorted alphabetically by key name. Each tag is a key-value pair. For a gateway with more than 10 tags assigned, you can view all tags using the ListTagsForResource API operation.

sMBFileShareInfo_accessBasedEnumeration :: Lens' SMBFileShareInfo (Maybe Bool)

Indicates whether AccessBasedEnumeration is enabled.

sMBFileShareInfo_adminUserList :: Lens' SMBFileShareInfo (Maybe [Text])

A list of users or groups in the Active Directory that have administrator rights to the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

sMBFileShareInfo_auditDestinationARN :: Lens' SMBFileShareInfo (Maybe Text)

The Amazon Resource Name (ARN) of the storage used for audit logs.

sMBFileShareInfo_fileShareStatus :: Lens' SMBFileShareInfo (Maybe Text)

Undocumented member.

sMBFileShareInfo_invalidUserList :: Lens' SMBFileShareInfo (Maybe [Text])

A list of users or groups in the Active Directory that are not allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

sMBFileShareInfo_kmsKey :: Lens' SMBFileShareInfo (Maybe Text)

Undocumented member.

sMBFileShareInfo_validUserList :: Lens' SMBFileShareInfo (Maybe [Text])

A list of users or groups in the Active Directory that are allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

sMBFileShareInfo_gatewayARN :: Lens' SMBFileShareInfo (Maybe Text)

Undocumented member.

sMBFileShareInfo_path :: Lens' SMBFileShareInfo (Maybe Text)

The file share path used by the SMB client to identify the mount point.

sMBFileShareInfo_vPCEndpointDNSName :: Lens' SMBFileShareInfo (Maybe Text)

Specifies the DNS name for the VPC endpoint that the SMB file share uses to connect to Amazon S3.

This parameter is required for SMB file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

sMBFileShareInfo_authentication :: Lens' SMBFileShareInfo (Maybe Text)

Undocumented member.

sMBFileShareInfo_cacheAttributes :: Lens' SMBFileShareInfo (Maybe CacheAttributes)

Refresh cache information for the file share.

sMBFileShareInfo_objectACL :: Lens' SMBFileShareInfo (Maybe ObjectACL)

Undocumented member.

sMBFileShareInfo_kmsEncrypted :: Lens' SMBFileShareInfo (Maybe Bool)

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

sMBFileShareInfo_fileShareId :: Lens' SMBFileShareInfo (Maybe Text)

Undocumented member.

sMBFileShareInfo_fileShareARN :: Lens' SMBFileShareInfo (Maybe Text)

Undocumented member.

sMBFileShareInfo_defaultStorageClass :: Lens' SMBFileShareInfo (Maybe Text)

The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

sMBFileShareInfo_fileShareName :: Lens' SMBFileShareInfo (Maybe Text)

The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN.

sMBFileShareInfo_role :: Lens' SMBFileShareInfo (Maybe Text)

Undocumented member.

sMBFileShareInfo_sMBACLEnabled :: Lens' SMBFileShareInfo (Maybe Bool)

If this value is set to true, it indicates that access control list (ACL) is enabled on the SMB file share. If it is set to false, it indicates that file and directory permissions are mapped to the POSIX permission.

For more information, see Using Microsoft Windows ACLs to control access to an SMB file share in the Storage Gateway User Guide.

sMBFileShareInfo_oplocksEnabled :: Lens' SMBFileShareInfo (Maybe Bool)

Specifies whether opportunistic locking is enabled for the SMB file share.

Enabling opportunistic locking on case-sensitive shares is not recommended for workloads that involve access to files with the same name in different case.

Valid Values: true | false

sMBFileShareInfo_notificationPolicy :: Lens' SMBFileShareInfo (Maybe Text)

The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

sMBFileShareInfo_requesterPays :: Lens' SMBFileShareInfo (Maybe Bool)

A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

sMBFileShareInfo_locationARN :: Lens' SMBFileShareInfo (Maybe Text)

Undocumented member.

sMBFileShareInfo_guessMIMETypeEnabled :: Lens' SMBFileShareInfo (Maybe Bool)

A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

sMBFileShareInfo_readOnly :: Lens' SMBFileShareInfo (Maybe Bool)

A value that sets the write status of a file share. Set this value to true to set the write status to read-only, otherwise set to false.

Valid Values: true | false

sMBFileShareInfo_bucketRegion :: Lens' SMBFileShareInfo (Maybe Text)

Specifies the Region of the S3 bucket where the SMB file share stores files.

This parameter is required for SMB file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

sMBFileShareInfo_caseSensitivity :: Lens' SMBFileShareInfo (Maybe CaseSensitivity)

The case of an object name in an Amazon S3 bucket. For ClientSpecified, the client determines the case sensitivity. For CaseSensitive, the gateway determines the case sensitivity. The default value is ClientSpecified.

sMBFileShareInfo_tags :: Lens' SMBFileShareInfo (Maybe [Tag])

A list of up to 50 tags assigned to the SMB file share, sorted alphabetically by key name. Each tag is a key-value pair. For a gateway with more than 10 tags assigned, you can view all tags using the ListTagsForResource API operation.