Network.AWS.StorageGateway.Types.Tape

Description

Documentation

data Tape

Describes a virtual tape object.

See: newTape smart constructor.

Constructors

Tape'

Fields

tapeBarcode :: Maybe Text

The barcode that identifies a specific virtual tape.

tapeStatus :: Maybe Text

The current state of the virtual tape.

kmsKey :: Maybe Text

tapeARN :: Maybe Text

The Amazon Resource Name (ARN) of the virtual tape.

progress :: Maybe Double

For archiving virtual tapes, indicates how much data remains to be uploaded before archiving is complete.

Range: 0 (not started) to 100 (complete).

tapeSizeInBytes :: Maybe Integer

The size, in bytes, of the virtual tape capacity.

vTLDevice :: Maybe Text

The virtual tape library (VTL) device that the virtual tape is associated with.

poolId :: Maybe Text

The ID of the pool that contains tapes that will be archived. The tapes in this pool are archived in the S3 storage class that is associated with the pool. When you use your backup application to eject the tape, the tape is archived directly into the storage class (S3 Glacier or S3 Glacier Deep Archive) that corresponds to the pool.

Valid Values: GLACIER | DEEP_ARCHIVE

tapeUsedInBytes :: Maybe Integer

The size, in bytes, of data stored on the virtual tape.

This value is not available for tapes created prior to May 13, 2015.

tapeCreatedDate :: Maybe POSIX

The date the virtual tape was created.

poolEntryDate :: Maybe POSIX

The date that the tape enters a custom tape pool.

worm :: Maybe Bool

If the tape is archived as write-once-read-many (WORM), this value is true.

retentionStartDate :: Maybe POSIX

The date that the tape is first archived with tape retention lock enabled.

newTape :: Tape

Create a value of Tape with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tapeBarcode:Tape', tape_tapeBarcode - The barcode that identifies a specific virtual tape.

$sel:tapeStatus:Tape', tape_tapeStatus - The current state of the virtual tape.

$sel:kmsKey:Tape', tape_kmsKey - Undocumented member.

$sel:tapeARN:Tape', tape_tapeARN - The Amazon Resource Name (ARN) of the virtual tape.

$sel:progress:Tape', tape_progress - For archiving virtual tapes, indicates how much data remains to be uploaded before archiving is complete.

Range: 0 (not started) to 100 (complete).

$sel:tapeSizeInBytes:Tape', tape_tapeSizeInBytes - The size, in bytes, of the virtual tape capacity.

$sel:vTLDevice:Tape', tape_vTLDevice - The virtual tape library (VTL) device that the virtual tape is associated with.

$sel:poolId:Tape', tape_poolId - The ID of the pool that contains tapes that will be archived. The tapes in this pool are archived in the S3 storage class that is associated with the pool. When you use your backup application to eject the tape, the tape is archived directly into the storage class (S3 Glacier or S3 Glacier Deep Archive) that corresponds to the pool.

Valid Values: GLACIER | DEEP_ARCHIVE

$sel:tapeUsedInBytes:Tape', tape_tapeUsedInBytes - The size, in bytes, of data stored on the virtual tape.

This value is not available for tapes created prior to May 13, 2015.

$sel:tapeCreatedDate:Tape', tape_tapeCreatedDate - The date the virtual tape was created.

$sel:poolEntryDate:Tape', tape_poolEntryDate - The date that the tape enters a custom tape pool.

$sel:worm:Tape', tape_worm - If the tape is archived as write-once-read-many (WORM), this value is true.

$sel:retentionStartDate:Tape', tape_retentionStartDate - The date that the tape is first archived with tape retention lock enabled.

tape_tapeBarcode :: Lens' Tape (Maybe Text)

The barcode that identifies a specific virtual tape.

tape_tapeStatus :: Lens' Tape (Maybe Text)

The current state of the virtual tape.

tape_kmsKey :: Lens' Tape (Maybe Text)

Undocumented member.

tape_tapeARN :: Lens' Tape (Maybe Text)

The Amazon Resource Name (ARN) of the virtual tape.

tape_progress :: Lens' Tape (Maybe Double)

For archiving virtual tapes, indicates how much data remains to be uploaded before archiving is complete.

Range: 0 (not started) to 100 (complete).

tape_tapeSizeInBytes :: Lens' Tape (Maybe Integer)

The size, in bytes, of the virtual tape capacity.

tape_vTLDevice :: Lens' Tape (Maybe Text)

The virtual tape library (VTL) device that the virtual tape is associated with.

tape_poolId :: Lens' Tape (Maybe Text)

The ID of the pool that contains tapes that will be archived. The tapes in this pool are archived in the S3 storage class that is associated with the pool. When you use your backup application to eject the tape, the tape is archived directly into the storage class (S3 Glacier or S3 Glacier Deep Archive) that corresponds to the pool.

Valid Values: GLACIER | DEEP_ARCHIVE

tape_tapeUsedInBytes :: Lens' Tape (Maybe Integer)

The size, in bytes, of data stored on the virtual tape.

This value is not available for tapes created prior to May 13, 2015.

tape_tapeCreatedDate :: Lens' Tape (Maybe UTCTime)

The date the virtual tape was created.

tape_poolEntryDate :: Lens' Tape (Maybe UTCTime)

The date that the tape enters a custom tape pool.

tape_worm :: Lens' Tape (Maybe Bool)

If the tape is archived as write-once-read-many (WORM), this value is true.

tape_retentionStartDate :: Lens' Tape (Maybe UTCTime)

The date that the tape is first archived with tape retention lock enabled.