Network.AWS.StorageGateway.Types.TapeArchive

Description

Documentation

data TapeArchive

Represents a virtual tape that is archived in the virtual tape shelf (VTS).

See: newTapeArchive smart constructor.

Constructors

TapeArchive'

Fields

tapeBarcode :: Maybe Text

The barcode that identifies the archived virtual tape.

tapeStatus :: Maybe Text

The current state of the archived virtual tape.

kmsKey :: Maybe Text

tapeARN :: Maybe Text

The Amazon Resource Name (ARN) of an archived virtual tape.

tapeSizeInBytes :: Maybe Integer

The size, in bytes, of the archived virtual tape.

completionTime :: Maybe POSIX

The time that the archiving of the virtual tape was completed.

The default timestamp format is in the ISO8601 extended YYYY-MM-DD'T'HH:MM:SS'Z' format.

poolId :: Maybe Text

The ID of the pool that was used to archive the tape. The tapes in this pool are archived in the S3 storage class that is associated with the pool.

Valid Values: GLACIER | DEEP_ARCHIVE

tapeUsedInBytes :: Maybe Integer

The size, in bytes, of data stored on the virtual tape.

This value is not available for tapes created prior to May 13, 2015.

tapeCreatedDate :: Maybe POSIX

The date the virtual tape was created.

poolEntryDate :: Maybe POSIX

The time that the tape entered the custom tape pool.

The default timestamp format is in the ISO8601 extended YYYY-MM-DD'T'HH:MM:SS'Z' format.

worm :: Maybe Bool

Set to true if the archived tape is stored as write-once-read-many (WORM).

retentionStartDate :: Maybe POSIX

If the archived tape is subject to tape retention lock, the date that the archived tape started being retained.

retrievedTo :: Maybe Text

The Amazon Resource Name (ARN) of the tape gateway that the virtual tape is being retrieved to.

The virtual tape is retrieved from the virtual tape shelf (VTS).

newTapeArchive :: TapeArchive

Create a value of TapeArchive with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tapeBarcode:TapeArchive', tapeArchive_tapeBarcode - The barcode that identifies the archived virtual tape.

$sel:tapeStatus:TapeArchive', tapeArchive_tapeStatus - The current state of the archived virtual tape.

$sel:kmsKey:TapeArchive', tapeArchive_kmsKey - Undocumented member.

$sel:tapeARN:TapeArchive', tapeArchive_tapeARN - The Amazon Resource Name (ARN) of an archived virtual tape.

$sel:tapeSizeInBytes:TapeArchive', tapeArchive_tapeSizeInBytes - The size, in bytes, of the archived virtual tape.

$sel:completionTime:TapeArchive', tapeArchive_completionTime - The time that the archiving of the virtual tape was completed.

The default timestamp format is in the ISO8601 extended YYYY-MM-DD'T'HH:MM:SS'Z' format.

$sel:poolId:TapeArchive', tapeArchive_poolId - The ID of the pool that was used to archive the tape. The tapes in this pool are archived in the S3 storage class that is associated with the pool.

Valid Values: GLACIER | DEEP_ARCHIVE

$sel:tapeUsedInBytes:TapeArchive', tapeArchive_tapeUsedInBytes - The size, in bytes, of data stored on the virtual tape.

This value is not available for tapes created prior to May 13, 2015.

$sel:tapeCreatedDate:TapeArchive', tapeArchive_tapeCreatedDate - The date the virtual tape was created.

$sel:poolEntryDate:TapeArchive', tapeArchive_poolEntryDate - The time that the tape entered the custom tape pool.

The default timestamp format is in the ISO8601 extended YYYY-MM-DD'T'HH:MM:SS'Z' format.

$sel:worm:TapeArchive', tapeArchive_worm - Set to true if the archived tape is stored as write-once-read-many (WORM).

$sel:retentionStartDate:TapeArchive', tapeArchive_retentionStartDate - If the archived tape is subject to tape retention lock, the date that the archived tape started being retained.

$sel:retrievedTo:TapeArchive', tapeArchive_retrievedTo - The Amazon Resource Name (ARN) of the tape gateway that the virtual tape is being retrieved to.

The virtual tape is retrieved from the virtual tape shelf (VTS).

tapeArchive_tapeBarcode :: Lens' TapeArchive (Maybe Text)

The barcode that identifies the archived virtual tape.

tapeArchive_tapeStatus :: Lens' TapeArchive (Maybe Text)

The current state of the archived virtual tape.

tapeArchive_kmsKey :: Lens' TapeArchive (Maybe Text)

Undocumented member.

tapeArchive_tapeARN :: Lens' TapeArchive (Maybe Text)

The Amazon Resource Name (ARN) of an archived virtual tape.

tapeArchive_tapeSizeInBytes :: Lens' TapeArchive (Maybe Integer)

The size, in bytes, of the archived virtual tape.

tapeArchive_completionTime :: Lens' TapeArchive (Maybe UTCTime)

The time that the archiving of the virtual tape was completed.

The default timestamp format is in the ISO8601 extended YYYY-MM-DD'T'HH:MM:SS'Z' format.

tapeArchive_poolId :: Lens' TapeArchive (Maybe Text)

The ID of the pool that was used to archive the tape. The tapes in this pool are archived in the S3 storage class that is associated with the pool.

Valid Values: GLACIER | DEEP_ARCHIVE

tapeArchive_tapeUsedInBytes :: Lens' TapeArchive (Maybe Integer)

The size, in bytes, of data stored on the virtual tape.

This value is not available for tapes created prior to May 13, 2015.

tapeArchive_tapeCreatedDate :: Lens' TapeArchive (Maybe UTCTime)

The date the virtual tape was created.

tapeArchive_poolEntryDate :: Lens' TapeArchive (Maybe UTCTime)

The time that the tape entered the custom tape pool.

The default timestamp format is in the ISO8601 extended YYYY-MM-DD'T'HH:MM:SS'Z' format.

tapeArchive_worm :: Lens' TapeArchive (Maybe Bool)

Set to true if the archived tape is stored as write-once-read-many (WORM).

tapeArchive_retentionStartDate :: Lens' TapeArchive (Maybe UTCTime)

If the archived tape is subject to tape retention lock, the date that the archived tape started being retained.

tapeArchive_retrievedTo :: Lens' TapeArchive (Maybe Text)

The Amazon Resource Name (ARN) of the tape gateway that the virtual tape is being retrieved to.

The virtual tape is retrieved from the virtual tape shelf (VTS).