Network.AWS.StorageGateway.Types.TapeInfo

Description

Documentation

data TapeInfo

Describes a virtual tape.

See: newTapeInfo smart constructor.

Constructors

TapeInfo'

Fields

tapeBarcode :: Maybe Text

The barcode that identifies a specific virtual tape.

tapeStatus :: Maybe Text

The status of the tape.

tapeARN :: Maybe Text

The Amazon Resource Name (ARN) of a virtual tape.

gatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and Region.

tapeSizeInBytes :: Maybe Integer

The size, in bytes, of a virtual tape.

poolId :: Maybe Text

The ID of the pool that you want to add your tape to for archiving. The tape in this pool is archived in the S3 storage class that is associated with the pool. When you use your backup application to eject the tape, the tape is archived directly into the storage class (S3 Glacier or S3 Glacier Deep Archive) that corresponds to the pool.

Valid Values: GLACIER | DEEP_ARCHIVE

poolEntryDate :: Maybe POSIX

The date that the tape entered the custom tape pool with tape retention lock enabled.

retentionStartDate :: Maybe POSIX

The date that the tape became subject to tape retention lock.

newTapeInfo :: TapeInfo

Create a value of TapeInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tapeBarcode:TapeInfo', tapeInfo_tapeBarcode - The barcode that identifies a specific virtual tape.

$sel:tapeStatus:TapeInfo', tapeInfo_tapeStatus - The status of the tape.

$sel:tapeARN:TapeInfo', tapeInfo_tapeARN - The Amazon Resource Name (ARN) of a virtual tape.

$sel:gatewayARN:TapeInfo', tapeInfo_gatewayARN - The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and Region.

$sel:tapeSizeInBytes:TapeInfo', tapeInfo_tapeSizeInBytes - The size, in bytes, of a virtual tape.

$sel:poolId:TapeInfo', tapeInfo_poolId - The ID of the pool that you want to add your tape to for archiving. The tape in this pool is archived in the S3 storage class that is associated with the pool. When you use your backup application to eject the tape, the tape is archived directly into the storage class (S3 Glacier or S3 Glacier Deep Archive) that corresponds to the pool.

Valid Values: GLACIER | DEEP_ARCHIVE

$sel:poolEntryDate:TapeInfo', tapeInfo_poolEntryDate - The date that the tape entered the custom tape pool with tape retention lock enabled.

$sel:retentionStartDate:TapeInfo', tapeInfo_retentionStartDate - The date that the tape became subject to tape retention lock.

tapeInfo_tapeBarcode :: Lens' TapeInfo (Maybe Text)

The barcode that identifies a specific virtual tape.

tapeInfo_tapeStatus :: Lens' TapeInfo (Maybe Text)

The status of the tape.

tapeInfo_tapeARN :: Lens' TapeInfo (Maybe Text)

The Amazon Resource Name (ARN) of a virtual tape.

tapeInfo_gatewayARN :: Lens' TapeInfo (Maybe Text)

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and Region.

tapeInfo_tapeSizeInBytes :: Lens' TapeInfo (Maybe Integer)

The size, in bytes, of a virtual tape.

tapeInfo_poolId :: Lens' TapeInfo (Maybe Text)

The ID of the pool that you want to add your tape to for archiving. The tape in this pool is archived in the S3 storage class that is associated with the pool. When you use your backup application to eject the tape, the tape is archived directly into the storage class (S3 Glacier or S3 Glacier Deep Archive) that corresponds to the pool.

Valid Values: GLACIER | DEEP_ARCHIVE

tapeInfo_poolEntryDate :: Lens' TapeInfo (Maybe UTCTime)

The date that the tape entered the custom tape pool with tape retention lock enabled.

tapeInfo_retentionStartDate :: Lens' TapeInfo (Maybe UTCTime)

The date that the tape became subject to tape retention lock.