Network.AWS.StorageGateway.Types.VTLDevice

Description

Documentation

data VTLDevice

Represents a device object associated with a tape gateway.

See: newVTLDevice smart constructor.

Constructors

VTLDevice'

Fields

deviceiSCSIAttributes :: Maybe DeviceiSCSIAttributes

A list of iSCSI information about a VTL device.

vTLDeviceVendor :: Maybe Text

Specifies the vendor of the device that the VTL device object emulates.

vTLDeviceARN :: Maybe Text

Specifies the unique Amazon Resource Name (ARN) of the device (tape drive or media changer).

vTLDeviceType :: Maybe Text

Specifies the type of device that the VTL device emulates.

vTLDeviceProductIdentifier :: Maybe Text

Specifies the model number of device that the VTL device emulates.

newVTLDevice :: VTLDevice

Create a value of VTLDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceiSCSIAttributes:VTLDevice', vTLDevice_deviceiSCSIAttributes - A list of iSCSI information about a VTL device.

$sel:vTLDeviceVendor:VTLDevice', vTLDevice_vTLDeviceVendor - Specifies the vendor of the device that the VTL device object emulates.

$sel:vTLDeviceARN:VTLDevice', vTLDevice_vTLDeviceARN - Specifies the unique Amazon Resource Name (ARN) of the device (tape drive or media changer).

$sel:vTLDeviceType:VTLDevice', vTLDevice_vTLDeviceType - Specifies the type of device that the VTL device emulates.

$sel:vTLDeviceProductIdentifier:VTLDevice', vTLDevice_vTLDeviceProductIdentifier - Specifies the model number of device that the VTL device emulates.

vTLDevice_deviceiSCSIAttributes :: Lens' VTLDevice (Maybe DeviceiSCSIAttributes)

A list of iSCSI information about a VTL device.

vTLDevice_vTLDeviceVendor :: Lens' VTLDevice (Maybe Text)

Specifies the vendor of the device that the VTL device object emulates.

vTLDevice_vTLDeviceARN :: Lens' VTLDevice (Maybe Text)

Specifies the unique Amazon Resource Name (ARN) of the device (tape drive or media changer).

vTLDevice_vTLDeviceType :: Lens' VTLDevice (Maybe Text)

Specifies the type of device that the VTL device emulates.

vTLDevice_vTLDeviceProductIdentifier :: Lens' VTLDevice (Maybe Text)

Specifies the model number of device that the VTL device emulates.