Network.AWS.StorageGateway.Types.VolumeRecoveryPointInfo

Description

Documentation

data VolumeRecoveryPointInfo

Describes a storage volume recovery point object.

See: newVolumeRecoveryPointInfo smart constructor.

Constructors

VolumeRecoveryPointInfo'

Fields

volumeRecoveryPointTime :: Maybe Text

The time the recovery point was taken.

volumeARN :: Maybe Text

The Amazon Resource Name (ARN) of the volume target.

volumeSizeInBytes :: Maybe Integer

The size of the volume in bytes.

volumeUsageInBytes :: Maybe Integer

The size of the data stored on the volume in bytes.

This value is not available for volumes created prior to May 13, 2015, until you store data on the volume.

newVolumeRecoveryPointInfo :: VolumeRecoveryPointInfo

Create a value of VolumeRecoveryPointInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:volumeRecoveryPointTime:VolumeRecoveryPointInfo', volumeRecoveryPointInfo_volumeRecoveryPointTime - The time the recovery point was taken.

$sel:volumeARN:VolumeRecoveryPointInfo', volumeRecoveryPointInfo_volumeARN - The Amazon Resource Name (ARN) of the volume target.

$sel:volumeSizeInBytes:VolumeRecoveryPointInfo', volumeRecoveryPointInfo_volumeSizeInBytes - The size of the volume in bytes.

$sel:volumeUsageInBytes:VolumeRecoveryPointInfo', volumeRecoveryPointInfo_volumeUsageInBytes - The size of the data stored on the volume in bytes.

This value is not available for volumes created prior to May 13, 2015, until you store data on the volume.

volumeRecoveryPointInfo_volumeRecoveryPointTime :: Lens' VolumeRecoveryPointInfo (Maybe Text)

The time the recovery point was taken.

volumeRecoveryPointInfo_volumeARN :: Lens' VolumeRecoveryPointInfo (Maybe Text)

The Amazon Resource Name (ARN) of the volume target.

volumeRecoveryPointInfo_volumeSizeInBytes :: Lens' VolumeRecoveryPointInfo (Maybe Integer)

The size of the volume in bytes.

volumeRecoveryPointInfo_volumeUsageInBytes :: Lens' VolumeRecoveryPointInfo (Maybe Integer)

The size of the data stored on the volume in bytes.

This value is not available for volumes created prior to May 13, 2015, until you store data on the volume.