Network.AWS.StorageGateway.UpdateGatewayInformation

Description

Updates a gateway's metadata, which includes the gateway's name and time zone. To specify which gateway to update, use the Amazon Resource Name (ARN) of the gateway in your request.

For gateways activated after September 2, 2015, the gateway's ARN contains the gateway ID rather than the gateway name. However, changing the name of the gateway has no effect on the gateway's ARN.

Creating a Request

data UpdateGatewayInformation

See: newUpdateGatewayInformation smart constructor.

Constructors

UpdateGatewayInformation'

Fields

gatewayCapacity :: Maybe GatewayCapacity

Specifies the size of the gateway's metadata cache.

gatewayName :: Maybe Text

gatewayTimezone :: Maybe Text

A value that indicates the time zone of the gateway.

cloudWatchLogGroupARN :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon CloudWatch log group that you want to use to monitor and log events in the gateway.

For more information, see What is Amazon CloudWatch Logs?

gatewayARN :: Text

newUpdateGatewayInformation

Arguments

:: Text
$sel:gatewayARN:UpdateGatewayInformation'

-> UpdateGatewayInformation

Create a value of UpdateGatewayInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayCapacity:UpdateGatewayInformation', updateGatewayInformation_gatewayCapacity - Specifies the size of the gateway's metadata cache.

$sel:gatewayName:UpdateGatewayInformation', updateGatewayInformation_gatewayName - Undocumented member.

$sel:gatewayTimezone:UpdateGatewayInformation', updateGatewayInformation_gatewayTimezone - A value that indicates the time zone of the gateway.

$sel:cloudWatchLogGroupARN:UpdateGatewayInformation', updateGatewayInformation_cloudWatchLogGroupARN - The Amazon Resource Name (ARN) of the Amazon CloudWatch log group that you want to use to monitor and log events in the gateway.

For more information, see What is Amazon CloudWatch Logs?

$sel:gatewayARN:UpdateGatewayInformation', updateGatewayInformation_gatewayARN - Undocumented member.

Request Lenses

updateGatewayInformation_gatewayCapacity :: Lens' UpdateGatewayInformation (Maybe GatewayCapacity)

Specifies the size of the gateway's metadata cache.

updateGatewayInformation_gatewayName :: Lens' UpdateGatewayInformation (Maybe Text)

Undocumented member.

updateGatewayInformation_gatewayTimezone :: Lens' UpdateGatewayInformation (Maybe Text)

A value that indicates the time zone of the gateway.

updateGatewayInformation_cloudWatchLogGroupARN :: Lens' UpdateGatewayInformation (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon CloudWatch log group that you want to use to monitor and log events in the gateway.

For more information, see What is Amazon CloudWatch Logs?

updateGatewayInformation_gatewayARN :: Lens' UpdateGatewayInformation Text

Undocumented member.

Destructuring the Response

data UpdateGatewayInformationResponse

A JSON object containing the Amazon Resource Name (ARN) of the gateway that was updated.

See: newUpdateGatewayInformationResponse smart constructor.

Constructors

UpdateGatewayInformationResponse'

Fields

gatewayARN :: Maybe Text

gatewayName :: Maybe Text

The name you configured for your gateway.

httpStatus :: Int

The response's http status code.

newUpdateGatewayInformationResponse

Arguments

:: Int
$sel:httpStatus:UpdateGatewayInformationResponse'

-> UpdateGatewayInformationResponse

Create a value of UpdateGatewayInformationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:UpdateGatewayInformation', updateGatewayInformationResponse_gatewayARN - Undocumented member.

$sel:gatewayName:UpdateGatewayInformation', updateGatewayInformationResponse_gatewayName - The name you configured for your gateway.

$sel:httpStatus:UpdateGatewayInformationResponse', updateGatewayInformationResponse_httpStatus - The response's http status code.

Response Lenses

updateGatewayInformationResponse_gatewayARN :: Lens' UpdateGatewayInformationResponse (Maybe Text)

Undocumented member.

updateGatewayInformationResponse_gatewayName :: Lens' UpdateGatewayInformationResponse (Maybe Text)

The name you configured for your gateway.

updateGatewayInformationResponse_httpStatus :: Lens' UpdateGatewayInformationResponse Int

The response's http status code.