Network.AWS.StorageGateway.UpdateMaintenanceStartTime

Description

Updates a gateway's weekly maintenance start time information, including day and time of the week. The maintenance time is the time in your gateway's time zone.

Creating a Request

data UpdateMaintenanceStartTime

A JSON object containing the following fields:

UpdateMaintenanceStartTimeInput$DayOfMonth

UpdateMaintenanceStartTimeInput$DayOfWeek

UpdateMaintenanceStartTimeInput$HourOfDay

UpdateMaintenanceStartTimeInput$MinuteOfHour

See: newUpdateMaintenanceStartTime smart constructor.

Constructors

UpdateMaintenanceStartTime'

Fields

dayOfMonth :: Maybe Natural

The day of the month component of the maintenance start time represented as an ordinal number from 1 to 28, where 1 represents the first day of the month and 28 represents the last day of the month.

dayOfWeek :: Maybe Natural

The day of the week component of the maintenance start time week represented as an ordinal number from 0 to 6, where 0 represents Sunday and 6 Saturday.

gatewayARN :: Text

hourOfDay :: Natural

The hour component of the maintenance start time represented as hh, where hh is the hour (00 to 23). The hour of the day is in the time zone of the gateway.

minuteOfHour :: Natural

The minute component of the maintenance start time represented as mm, where mm is the minute (00 to 59). The minute of the hour is in the time zone of the gateway.

newUpdateMaintenanceStartTime

Arguments

:: Text
$sel:gatewayARN:UpdateMaintenanceStartTime'

-> Natural
$sel:hourOfDay:UpdateMaintenanceStartTime'

-> Natural
$sel:minuteOfHour:UpdateMaintenanceStartTime'

-> UpdateMaintenanceStartTime

Create a value of UpdateMaintenanceStartTime with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dayOfMonth:UpdateMaintenanceStartTime', updateMaintenanceStartTime_dayOfMonth - The day of the month component of the maintenance start time represented as an ordinal number from 1 to 28, where 1 represents the first day of the month and 28 represents the last day of the month.

$sel:dayOfWeek:UpdateMaintenanceStartTime', updateMaintenanceStartTime_dayOfWeek - The day of the week component of the maintenance start time week represented as an ordinal number from 0 to 6, where 0 represents Sunday and 6 Saturday.

$sel:gatewayARN:UpdateMaintenanceStartTime', updateMaintenanceStartTime_gatewayARN - Undocumented member.

$sel:hourOfDay:UpdateMaintenanceStartTime', updateMaintenanceStartTime_hourOfDay - The hour component of the maintenance start time represented as hh, where hh is the hour (00 to 23). The hour of the day is in the time zone of the gateway.

$sel:minuteOfHour:UpdateMaintenanceStartTime', updateMaintenanceStartTime_minuteOfHour - The minute component of the maintenance start time represented as mm, where mm is the minute (00 to 59). The minute of the hour is in the time zone of the gateway.

Request Lenses

updateMaintenanceStartTime_dayOfMonth :: Lens' UpdateMaintenanceStartTime (Maybe Natural)

The day of the month component of the maintenance start time represented as an ordinal number from 1 to 28, where 1 represents the first day of the month and 28 represents the last day of the month.

updateMaintenanceStartTime_dayOfWeek :: Lens' UpdateMaintenanceStartTime (Maybe Natural)

The day of the week component of the maintenance start time week represented as an ordinal number from 0 to 6, where 0 represents Sunday and 6 Saturday.

updateMaintenanceStartTime_gatewayARN :: Lens' UpdateMaintenanceStartTime Text

Undocumented member.

updateMaintenanceStartTime_hourOfDay :: Lens' UpdateMaintenanceStartTime Natural

The hour component of the maintenance start time represented as hh, where hh is the hour (00 to 23). The hour of the day is in the time zone of the gateway.

updateMaintenanceStartTime_minuteOfHour :: Lens' UpdateMaintenanceStartTime Natural

The minute component of the maintenance start time represented as mm, where mm is the minute (00 to 59). The minute of the hour is in the time zone of the gateway.

Destructuring the Response

data UpdateMaintenanceStartTimeResponse

A JSON object containing the Amazon Resource Name (ARN) of the gateway whose maintenance start time is updated.

See: newUpdateMaintenanceStartTimeResponse smart constructor.

Constructors

UpdateMaintenanceStartTimeResponse'

Fields

gatewayARN :: Maybe Text

httpStatus :: Int

The response's http status code.

newUpdateMaintenanceStartTimeResponse

Arguments

:: Int
$sel:httpStatus:UpdateMaintenanceStartTimeResponse'

-> UpdateMaintenanceStartTimeResponse

Create a value of UpdateMaintenanceStartTimeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:UpdateMaintenanceStartTime', updateMaintenanceStartTimeResponse_gatewayARN - Undocumented member.

$sel:httpStatus:UpdateMaintenanceStartTimeResponse', updateMaintenanceStartTimeResponse_httpStatus - The response's http status code.

Response Lenses

updateMaintenanceStartTimeResponse_gatewayARN :: Lens' UpdateMaintenanceStartTimeResponse (Maybe Text)

Undocumented member.

updateMaintenanceStartTimeResponse_httpStatus :: Lens' UpdateMaintenanceStartTimeResponse Int

The response's http status code.