Network.AWS.StorageGateway.UpdateSMBFileShare

Description

Updates a Server Message Block (SMB) file share. This operation is only supported for S3 File Gateways.

To leave a file share field unchanged, set the corresponding input field to null.

File gateways require Security Token Service (STS) to be activated to enable you to create a file share. Make sure that STS is activated in the Region you are creating your file gateway in. If STS is not activated in this Region, activate it. For information about how to activate STS, see Activating and deactivating STS in an Region in the Identity and Access Management User Guide.

File gateways don't support creating hard or symbolic links on a file share.

Creating a Request

data UpdateSMBFileShare

UpdateSMBFileShareInput

See: newUpdateSMBFileShare smart constructor.

Constructors

UpdateSMBFileShare'

Fields

accessBasedEnumeration :: Maybe Bool

The files and folders on this share will only be visible to users with read access.

adminUserList :: Maybe [Text]

A list of users or groups in the Active Directory that have administrator rights to the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

auditDestinationARN :: Maybe Text

The Amazon Resource Name (ARN) of the storage used for audit logs.

invalidUserList :: Maybe [Text]

A list of users or groups in the Active Directory that are not allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

kmsKey :: Maybe Text

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

validUserList :: Maybe [Text]

A list of users or groups in the Active Directory that are allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

cacheAttributes :: Maybe CacheAttributes

Specifies refresh cache information for the file share.

objectACL :: Maybe ObjectACL

A value that sets the access control list (ACL) permission for objects in the S3 bucket that a S3 File Gateway puts objects into. The default value is private.

kmsEncrypted :: Maybe Bool

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

defaultStorageClass :: Maybe Text

The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

fileShareName :: Maybe Text

The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN.

sMBACLEnabled :: Maybe Bool

Set this value to true to enable access control list (ACL) on the SMB file share. Set it to false to map file and directory permissions to the POSIX permissions.

For more information, see Using Microsoft Windows ACLs to control access to an SMB file share in the Storage Gateway User Guide.

Valid Values: true | false

oplocksEnabled :: Maybe Bool

Specifies whether opportunistic locking is enabled for the SMB file share.

Enabling opportunistic locking on case-sensitive shares is not recommended for workloads that involve access to files with the same name in different case.

Valid Values: true | false

notificationPolicy :: Maybe Text

The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

requesterPays :: Maybe Bool

A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

guessMIMETypeEnabled :: Maybe Bool

A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

readOnly :: Maybe Bool

A value that sets the write status of a file share. Set this value to true to set write status to read-only, otherwise set to false.

Valid Values: true | false

caseSensitivity :: Maybe CaseSensitivity

The case of an object name in an Amazon S3 bucket. For ClientSpecified, the client determines the case sensitivity. For CaseSensitive, the gateway determines the case sensitivity. The default value is ClientSpecified.

fileShareARN :: Text

The Amazon Resource Name (ARN) of the SMB file share that you want to update.

newUpdateSMBFileShare

Arguments

:: Text
$sel:fileShareARN:UpdateSMBFileShare'

-> UpdateSMBFileShare

Create a value of UpdateSMBFileShare with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessBasedEnumeration:UpdateSMBFileShare', updateSMBFileShare_accessBasedEnumeration - The files and folders on this share will only be visible to users with read access.

$sel:adminUserList:UpdateSMBFileShare', updateSMBFileShare_adminUserList - A list of users or groups in the Active Directory that have administrator rights to the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

$sel:auditDestinationARN:UpdateSMBFileShare', updateSMBFileShare_auditDestinationARN - The Amazon Resource Name (ARN) of the storage used for audit logs.

$sel:invalidUserList:UpdateSMBFileShare', updateSMBFileShare_invalidUserList - A list of users or groups in the Active Directory that are not allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

$sel:kmsKey:UpdateSMBFileShare', updateSMBFileShare_kmsKey - The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

$sel:validUserList:UpdateSMBFileShare', updateSMBFileShare_validUserList - A list of users or groups in the Active Directory that are allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

$sel:cacheAttributes:UpdateSMBFileShare', updateSMBFileShare_cacheAttributes - Specifies refresh cache information for the file share.

$sel:objectACL:UpdateSMBFileShare', updateSMBFileShare_objectACL - A value that sets the access control list (ACL) permission for objects in the S3 bucket that a S3 File Gateway puts objects into. The default value is private.

$sel:kmsEncrypted:UpdateSMBFileShare', updateSMBFileShare_kmsEncrypted - Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

$sel:defaultStorageClass:UpdateSMBFileShare', updateSMBFileShare_defaultStorageClass - The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

$sel:fileShareName:UpdateSMBFileShare', updateSMBFileShare_fileShareName - The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN.

$sel:sMBACLEnabled:UpdateSMBFileShare', updateSMBFileShare_sMBACLEnabled - Set this value to true to enable access control list (ACL) on the SMB file share. Set it to false to map file and directory permissions to the POSIX permissions.

For more information, see Using Microsoft Windows ACLs to control access to an SMB file share in the Storage Gateway User Guide.

Valid Values: true | false

$sel:oplocksEnabled:UpdateSMBFileShare', updateSMBFileShare_oplocksEnabled - Specifies whether opportunistic locking is enabled for the SMB file share.

Enabling opportunistic locking on case-sensitive shares is not recommended for workloads that involve access to files with the same name in different case.

Valid Values: true | false

$sel:notificationPolicy:UpdateSMBFileShare', updateSMBFileShare_notificationPolicy - The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

$sel:requesterPays:UpdateSMBFileShare', updateSMBFileShare_requesterPays - A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

$sel:guessMIMETypeEnabled:UpdateSMBFileShare', updateSMBFileShare_guessMIMETypeEnabled - A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

$sel:readOnly:UpdateSMBFileShare', updateSMBFileShare_readOnly - A value that sets the write status of a file share. Set this value to true to set write status to read-only, otherwise set to false.

Valid Values: true | false

$sel:caseSensitivity:UpdateSMBFileShare', updateSMBFileShare_caseSensitivity - The case of an object name in an Amazon S3 bucket. For ClientSpecified, the client determines the case sensitivity. For CaseSensitive, the gateway determines the case sensitivity. The default value is ClientSpecified.

$sel:fileShareARN:UpdateSMBFileShare', updateSMBFileShare_fileShareARN - The Amazon Resource Name (ARN) of the SMB file share that you want to update.

Request Lenses

updateSMBFileShare_accessBasedEnumeration :: Lens' UpdateSMBFileShare (Maybe Bool)

The files and folders on this share will only be visible to users with read access.

updateSMBFileShare_adminUserList :: Lens' UpdateSMBFileShare (Maybe [Text])

A list of users or groups in the Active Directory that have administrator rights to the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

updateSMBFileShare_auditDestinationARN :: Lens' UpdateSMBFileShare (Maybe Text)

The Amazon Resource Name (ARN) of the storage used for audit logs.

updateSMBFileShare_invalidUserList :: Lens' UpdateSMBFileShare (Maybe [Text])

A list of users or groups in the Active Directory that are not allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

updateSMBFileShare_kmsKey :: Lens' UpdateSMBFileShare (Maybe Text)

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

updateSMBFileShare_validUserList :: Lens' UpdateSMBFileShare (Maybe [Text])

A list of users or groups in the Active Directory that are allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

updateSMBFileShare_cacheAttributes :: Lens' UpdateSMBFileShare (Maybe CacheAttributes)

Specifies refresh cache information for the file share.

updateSMBFileShare_objectACL :: Lens' UpdateSMBFileShare (Maybe ObjectACL)

A value that sets the access control list (ACL) permission for objects in the S3 bucket that a S3 File Gateway puts objects into. The default value is private.

updateSMBFileShare_kmsEncrypted :: Lens' UpdateSMBFileShare (Maybe Bool)

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

updateSMBFileShare_defaultStorageClass :: Lens' UpdateSMBFileShare (Maybe Text)

The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

updateSMBFileShare_fileShareName :: Lens' UpdateSMBFileShare (Maybe Text)

The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN.

updateSMBFileShare_sMBACLEnabled :: Lens' UpdateSMBFileShare (Maybe Bool)

Set this value to true to enable access control list (ACL) on the SMB file share. Set it to false to map file and directory permissions to the POSIX permissions.

For more information, see Using Microsoft Windows ACLs to control access to an SMB file share in the Storage Gateway User Guide.

Valid Values: true | false

updateSMBFileShare_oplocksEnabled :: Lens' UpdateSMBFileShare (Maybe Bool)

Specifies whether opportunistic locking is enabled for the SMB file share.

Enabling opportunistic locking on case-sensitive shares is not recommended for workloads that involve access to files with the same name in different case.

Valid Values: true | false

updateSMBFileShare_notificationPolicy :: Lens' UpdateSMBFileShare (Maybe Text)

The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

updateSMBFileShare_requesterPays :: Lens' UpdateSMBFileShare (Maybe Bool)

A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

updateSMBFileShare_guessMIMETypeEnabled :: Lens' UpdateSMBFileShare (Maybe Bool)

A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

updateSMBFileShare_readOnly :: Lens' UpdateSMBFileShare (Maybe Bool)

A value that sets the write status of a file share. Set this value to true to set write status to read-only, otherwise set to false.

Valid Values: true | false

updateSMBFileShare_caseSensitivity :: Lens' UpdateSMBFileShare (Maybe CaseSensitivity)

The case of an object name in an Amazon S3 bucket. For ClientSpecified, the client determines the case sensitivity. For CaseSensitive, the gateway determines the case sensitivity. The default value is ClientSpecified.

updateSMBFileShare_fileShareARN :: Lens' UpdateSMBFileShare Text

The Amazon Resource Name (ARN) of the SMB file share that you want to update.

Destructuring the Response

data UpdateSMBFileShareResponse

UpdateSMBFileShareOutput

See: newUpdateSMBFileShareResponse smart constructor.

Constructors

UpdateSMBFileShareResponse'

Fields

fileShareARN :: Maybe Text

The Amazon Resource Name (ARN) of the updated SMB file share.

httpStatus :: Int

The response's http status code.

newUpdateSMBFileShareResponse

Arguments

:: Int
$sel:httpStatus:UpdateSMBFileShareResponse'

-> UpdateSMBFileShareResponse

Create a value of UpdateSMBFileShareResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fileShareARN:UpdateSMBFileShare', updateSMBFileShareResponse_fileShareARN - The Amazon Resource Name (ARN) of the updated SMB file share.

$sel:httpStatus:UpdateSMBFileShareResponse', updateSMBFileShareResponse_httpStatus - The response's http status code.

Response Lenses

updateSMBFileShareResponse_fileShareARN :: Lens' UpdateSMBFileShareResponse (Maybe Text)

The Amazon Resource Name (ARN) of the updated SMB file share.

updateSMBFileShareResponse_httpStatus :: Lens' UpdateSMBFileShareResponse Int

The response's http status code.