Network.AWS.StorageGateway.UpdateVTLDeviceType

Description

Updates the type of medium changer in a tape gateway. When you activate a tape gateway, you select a medium changer type for the tape gateway. This operation enables you to select a different type of medium changer after a tape gateway is activated. This operation is only supported in the tape gateway type.

Creating a Request

data UpdateVTLDeviceType

See: newUpdateVTLDeviceType smart constructor.

Constructors

UpdateVTLDeviceType'

Fields

vTLDeviceARN :: Text

The Amazon Resource Name (ARN) of the medium changer you want to select.

deviceType :: Text

The type of medium changer you want to select.

Valid Values: STK-L700 | AWS-Gateway-VTL | IBM-03584L32-0402

newUpdateVTLDeviceType

Arguments

:: Text
$sel:vTLDeviceARN:UpdateVTLDeviceType'

-> Text
$sel:deviceType:UpdateVTLDeviceType'

-> UpdateVTLDeviceType

Create a value of UpdateVTLDeviceType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vTLDeviceARN:UpdateVTLDeviceType', updateVTLDeviceType_vTLDeviceARN - The Amazon Resource Name (ARN) of the medium changer you want to select.

$sel:deviceType:UpdateVTLDeviceType', updateVTLDeviceType_deviceType - The type of medium changer you want to select.

Valid Values: STK-L700 | AWS-Gateway-VTL | IBM-03584L32-0402

Request Lenses

updateVTLDeviceType_vTLDeviceARN :: Lens' UpdateVTLDeviceType Text

The Amazon Resource Name (ARN) of the medium changer you want to select.

updateVTLDeviceType_deviceType :: Lens' UpdateVTLDeviceType Text

The type of medium changer you want to select.

Valid Values: STK-L700 | AWS-Gateway-VTL | IBM-03584L32-0402

Destructuring the Response

data UpdateVTLDeviceTypeResponse

UpdateVTLDeviceTypeOutput

See: newUpdateVTLDeviceTypeResponse smart constructor.

Constructors

UpdateVTLDeviceTypeResponse'

Fields

vTLDeviceARN :: Maybe Text

The Amazon Resource Name (ARN) of the medium changer you have selected.

httpStatus :: Int

The response's http status code.

newUpdateVTLDeviceTypeResponse

Arguments

:: Int
$sel:httpStatus:UpdateVTLDeviceTypeResponse'

-> UpdateVTLDeviceTypeResponse

Create a value of UpdateVTLDeviceTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vTLDeviceARN:UpdateVTLDeviceType', updateVTLDeviceTypeResponse_vTLDeviceARN - The Amazon Resource Name (ARN) of the medium changer you have selected.

$sel:httpStatus:UpdateVTLDeviceTypeResponse', updateVTLDeviceTypeResponse_httpStatus - The response's http status code.

Response Lenses

updateVTLDeviceTypeResponse_vTLDeviceARN :: Lens' UpdateVTLDeviceTypeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the medium changer you have selected.

updateVTLDeviceTypeResponse_httpStatus :: Lens' UpdateVTLDeviceTypeResponse Int

The response's http status code.