Network.AWS.StorageGateway.V2013_06_30

Description

AWS Storage Gateway is a service that connects an on-premises software appliance with cloud-based storage to provide seamless and secure integration between your on-premises IT environment and AWS's storage infrastructure.

Documentation

module Network.AWS.StorageGateway.V2013_06_30.ActivateGateway

module Network.AWS.StorageGateway.V2013_06_30.AddCache

module Network.AWS.StorageGateway.V2013_06_30.AddUploadBuffer

module Network.AWS.StorageGateway.V2013_06_30.AddWorkingStorage

module Network.AWS.StorageGateway.V2013_06_30.CancelArchival

module Network.AWS.StorageGateway.V2013_06_30.CancelRetrieval

module Network.AWS.StorageGateway.V2013_06_30.CreateCachediSCSIVolume

module Network.AWS.StorageGateway.V2013_06_30.CreateSnapshot

module Network.AWS.StorageGateway.V2013_06_30.CreateSnapshotFromVolumeRecoveryPoint

module Network.AWS.StorageGateway.V2013_06_30.CreateStorediSCSIVolume

module Network.AWS.StorageGateway.V2013_06_30.CreateTapes

module Network.AWS.StorageGateway.V2013_06_30.DeleteBandwidthRateLimit

module Network.AWS.StorageGateway.V2013_06_30.DeleteChapCredentials

module Network.AWS.StorageGateway.V2013_06_30.DeleteGateway

module Network.AWS.StorageGateway.V2013_06_30.DeleteSnapshotSchedule

module Network.AWS.StorageGateway.V2013_06_30.DeleteTape

module Network.AWS.StorageGateway.V2013_06_30.DeleteTapeArchive

module Network.AWS.StorageGateway.V2013_06_30.DeleteVolume

module Network.AWS.StorageGateway.V2013_06_30.DescribeBandwidthRateLimit

module Network.AWS.StorageGateway.V2013_06_30.DescribeCache

module Network.AWS.StorageGateway.V2013_06_30.DescribeCachediSCSIVolumes

module Network.AWS.StorageGateway.V2013_06_30.DescribeChapCredentials

module Network.AWS.StorageGateway.V2013_06_30.DescribeGatewayInformation

module Network.AWS.StorageGateway.V2013_06_30.DescribeMaintenanceStartTime

module Network.AWS.StorageGateway.V2013_06_30.DescribeSnapshotSchedule

module Network.AWS.StorageGateway.V2013_06_30.DescribeStorediSCSIVolumes

module Network.AWS.StorageGateway.V2013_06_30.DescribeTapeArchives

module Network.AWS.StorageGateway.V2013_06_30.DescribeTapeRecoveryPoints

module Network.AWS.StorageGateway.V2013_06_30.DescribeTapes

module Network.AWS.StorageGateway.V2013_06_30.DescribeUploadBuffer

module Network.AWS.StorageGateway.V2013_06_30.DescribeVTLDevices

module Network.AWS.StorageGateway.V2013_06_30.DescribeWorkingStorage

module Network.AWS.StorageGateway.V2013_06_30.DisableGateway

module Network.AWS.StorageGateway.V2013_06_30.ListGateways

module Network.AWS.StorageGateway.V2013_06_30.ListLocalDisks

module Network.AWS.StorageGateway.V2013_06_30.ListVolumeRecoveryPoints

module Network.AWS.StorageGateway.V2013_06_30.ListVolumes

module Network.AWS.StorageGateway.V2013_06_30.RetrieveTapeArchive

module Network.AWS.StorageGateway.V2013_06_30.RetrieveTapeRecoveryPoint

module Network.AWS.StorageGateway.V2013_06_30.ShutdownGateway

module Network.AWS.StorageGateway.V2013_06_30.StartGateway

module Network.AWS.StorageGateway.V2013_06_30.Types

module Network.AWS.StorageGateway.V2013_06_30.UpdateBandwidthRateLimit

module Network.AWS.StorageGateway.V2013_06_30.UpdateChapCredentials

module Network.AWS.StorageGateway.V2013_06_30.UpdateGatewayInformation

module Network.AWS.StorageGateway.V2013_06_30.UpdateGatewaySoftwareNow

module Network.AWS.StorageGateway.V2013_06_30.UpdateMaintenanceStartTime

module Network.AWS.StorageGateway.V2013_06_30.UpdateSnapshotSchedule