Network.AWS.StorageGateway.V2013_06_30.AddUploadBuffer

Description

This operation configures one or more gateway local disks as upload buffer for a specified gateway. This operation is supported for both the gateway-stored and gateway-cached volume architectures. In the request, you specify the gateway Amazon Resource Name (ARN) to which you want to add upload buffer, and one or more disk IDs that you want to configure as upload buffer.

Documentation

data AddUploadBuffer

Constructors

AddUploadBuffer

Fields

_aubiDiskIds :: [Text]

_aubiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

aubiGatewayARN :: Lens' AddUploadBuffer Text

aubiDiskIds :: Lens' AddUploadBuffer [Text]

data AddUploadBufferResponse

Constructors

AddUploadBufferResponse

Fields

_auboGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

auboGatewayARN :: Iso' AddUploadBufferResponse (Maybe Text)