Network.AWS.StorageGateway.V2013_06_30.DeleteBandwidthRateLimit

Description

This operation deletes the bandwidth rate limits of a gateway. You can delete either the upload and download bandwidth rate limit, or you can delete both. If you delete only one of the limits, the other limit remains unchanged. To specify which gateway to work with, use the Amazon Resource Name (ARN) of the gateway in your request. Example Request The following example shows a request that deletes both of the bandwidth rate limits of a gateway. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.DeleteBandwidthRateLimit { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway", "BandwidthType: All } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 80 { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" }.

Documentation

data DeleteBandwidthRateLimit

Constructors

DeleteBandwidthRateLimit

Fields

_dbrljBandwidthType :: Text

_dbrljGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dbrljGatewayARN :: Lens' DeleteBandwidthRateLimit Text

dbrljBandwidthType :: Lens' DeleteBandwidthRateLimit Text

data DeleteBandwidthRateLimitResponse

Constructors

DeleteBandwidthRateLimitResponse

Fields

_dbrlpGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dbrlpGatewayARN :: Iso' DeleteBandwidthRateLimitResponse (Maybe Text)