Network.AWS.StorageGateway.V2013_06_30.DeleteSnapshotSchedule

Description

This operation deletes a snapshot of a volume. You can take snapshots of your gateway volumes on a scheduled or ad-hoc basis. This API enables you to delete a snapshot schedule for a volume. For more information, see Working with Snapshots. In the DeleteSnapshotSchedule request, you identify the volume by providing its Amazon Resource Name (ARN). To list or delete a snapshot, you must use the Amazon EC2 API. in Amazon Elastic Compute Cloud API Reference. Example Request The following example... POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com Content-Type: application/x-amz-json-1.1 Authorization: AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE20120425us-east-1storagegatewayaws4_request, SignedHeaders=content-type;host;x-amz-date;x-amz-target, Signature=9cd5a3584d1d67d57e61f120f35102d6b3649066abdd4bf4bbcf05bd9f2f8fe2 x-amz-date: 20120912T120000Z x-amz-target: StorageGateway_20120630.DeleteSnapshotSchedule { VolumeARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewayvolume/vol-1122AABB" } HTTP/1.1 200 OK x-amzn-RequestId: gur28r2rqlgb8vvs0mq17hlgij1q8glle1qeu3kpgg6f0kstauu0 Date: Wed, 12 Sep 2012 12:00:02 GMT Content-Type: application/x-amz-json-1.1 Content-length: 137 { VolumeARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewayvolume/vol-1122AABB" }.

Documentation

data DeleteSnapshotSchedule

Constructors

DeleteSnapshotSchedule

Fields

_dssjVolumeARN :: Text

dssjVolumeARN :: Iso' DeleteSnapshotSchedule Text

data DeleteSnapshotScheduleResponse

Constructors

DeleteSnapshotScheduleResponse

Fields

_dsspVolumeARN :: Maybe Text

dsspVolumeARN :: Iso' DeleteSnapshotScheduleResponse (Maybe Text)