Network.AWS.StorageGateway.V2013_06_30.DescribeBandwidthRateLimit

Description

This operation returns the bandwidth rate limits of a gateway. By default, these limits are not set, which means no bandwidth rate limiting is in effect. This operation only returns a value for a bandwidth rate limit only if the limit is set. If no limits are set for the gateway, then this operation returns only the gateway ARN in the response body. To specify which gateway to describe, use the Amazon Resource Name (ARN) of the gateway in your request. Example Request The following example shows a request that returns the bandwidth throttle properties of a gateway. POST / HTTP/1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.DescribeBandwidthRateLimit { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygate way" } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 169 { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway", AverageUploadRateLimitInBitsPerSec: 102400, AverageDownloadRateLimitInBitsPerSec: 51200 }.

Documentation

data DescribeBandwidthRateLimit

Constructors

DescribeBandwidthRateLimit

Fields

_dbrliGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dbrliGatewayARN :: Iso' DescribeBandwidthRateLimit Text

data DescribeBandwidthRateLimitResponse

Constructors

DescribeBandwidthRateLimitResponse

Fields

_dbrloAverageDownloadRateLimitInBitsPerSec :: Maybe Integer

The average download bandwidth rate limit in bits per second. This field does not appear in the response if the download rate limit is not set.

_dbrloAverageUploadRateLimitInBitsPerSec :: Maybe Integer

The average upload bandwidth rate limit in bits per second. This field does not appear in the response if the upload rate limit is not set.

_dbrloGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dbrloGatewayARN :: Lens' DescribeBandwidthRateLimitResponse (Maybe Text)

dbrloAverageUploadRateLimitInBitsPerSec :: Lens' DescribeBandwidthRateLimitResponse (Maybe Integer)

dbrloAverageDownloadRateLimitInBitsPerSec :: Lens' DescribeBandwidthRateLimitResponse (Maybe Integer)