Network.AWS.StorageGateway.V2013_06_30.DescribeGatewayInformation

Description

This operation returns metadata about a gateway such as its name, network interfaces, configured time zone, and the state (whether the gateway is running or not). To specify which gateway to describe, use the Amazon Resource Name (ARN) of the gateway in your request. Example Request The following example shows a request for describing a gateway. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.DescribeGatewayInformation { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 227 { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway", GatewayId: "sgw-AABB1122", GatewayNetworkInterfaces: [ {Ipv4Address: "10.35.69.216"} ], GatewayState: STATE_RUNNING, GatewayTimezone: GMT-8:00 }.

Documentation

data DescribeGatewayInformation

Constructors

DescribeGatewayInformation

Fields

_dgiiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dgiiGatewayARN :: Iso' DescribeGatewayInformation Text

data DescribeGatewayInformationResponse

Constructors

DescribeGatewayInformationResponse

Fields

_dgioGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_dgioGatewayId :: Maybe Text

The gateway ID.

_dgioGatewayNetworkInterfaces :: [NetworkInterface]

A NetworkInterface array that contains descriptions of the gateway network interfaces.

_dgioGatewayState :: Maybe Text

One of the values that indicates the operating state of the gateway.

_dgioGatewayTimezone :: Maybe Text

One of the values that indicates the time zone configured for the gateway.

_dgioGatewayType :: Maybe Text

TBD.

_dgioNextUpdateAvailabilityDate :: Maybe Text

The date at which an update to the gateway is available. This date is in the time zone of the gateway. If the gateway is not available for an update this field is not returned in the response. response example. -->.

dgioNextUpdateAvailabilityDate :: Lens' DescribeGatewayInformationResponse (Maybe Text)

dgioGatewayType :: Lens' DescribeGatewayInformationResponse (Maybe Text)

dgioGatewayTimezone :: Lens' DescribeGatewayInformationResponse (Maybe Text)

dgioGatewayState :: Lens' DescribeGatewayInformationResponse (Maybe Text)

dgioGatewayNetworkInterfaces :: Lens' DescribeGatewayInformationResponse [NetworkInterface]

dgioGatewayId :: Lens' DescribeGatewayInformationResponse (Maybe Text)

dgioGatewayARN :: Lens' DescribeGatewayInformationResponse (Maybe Text)