Network.AWS.StorageGateway.V2013_06_30.DescribeMaintenanceStartTime

Description

This operation returns your gateway's weekly maintenance start time including the day and time of the week. Note that values are in terms of the gateway's time zone. Example Request The following example shows a request that describes a gateway's maintenance window. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.DescribeMaintenanceStartTime { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 136 { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway", HourOfDay: 15, MinuteOfHour: 35, DayOfWeek: 2, Timezone: "GMT+7:00" }.

Documentation

data DescribeMaintenanceStartTime

Constructors

DescribeMaintenanceStartTime

Fields

_dmstiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dmstiGatewayARN :: Iso' DescribeMaintenanceStartTime Text

data DescribeMaintenanceStartTimeResponse

Constructors

DescribeMaintenanceStartTimeResponse

Fields

_dmstoDayOfWeek :: Maybe Integer

_dmstoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_dmstoTimezone :: Maybe Text

_dmstoHourOfDay :: Maybe Integer

_dmstoMinuteOfHour :: Maybe Integer

dmstoTimezone :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Text)

dmstoMinuteOfHour :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Integer)

dmstoHourOfDay :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Integer)

dmstoGatewayARN :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Text)

dmstoDayOfWeek :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Integer)