Network.AWS.StorageGateway.V2013_06_30.DescribeVTLDevices

Documentation

describeVTLDevices

Arguments

:: Text
_dvtldiGatewayARN

-> DescribeVTLDevices

Minimum specification for a DescribeVTLDevices request.

data DescribeVTLDevices

Constructors

DescribeVTLDevices

Fields

_dvtldiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_dvtldiMarker :: Maybe Text

_dvtldiLimit :: Maybe Integer

_dvtldiVTLDeviceARNs :: [Text]

dvtldiVTLDeviceARNs :: Lens' DescribeVTLDevices [Text]

dvtldiMarker :: Lens' DescribeVTLDevices (Maybe Text)

dvtldiLimit :: Lens' DescribeVTLDevices (Maybe Integer)

dvtldiGatewayARN :: Lens' DescribeVTLDevices Text

data DescribeVTLDevicesResponse

Constructors

DescribeVTLDevicesResponse

Fields

_dvtldoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_dvtldoMarker :: Maybe Text

_dvtldoVTLDevices :: [VTLDevice]

dvtldoVTLDevices :: Lens' DescribeVTLDevicesResponse [VTLDevice]

dvtldoMarker :: Lens' DescribeVTLDevicesResponse (Maybe Text)

dvtldoGatewayARN :: Lens' DescribeVTLDevicesResponse (Maybe Text)