Network.AWS.StorageGateway.V2013_06_30.DisableGateway

Documentation

data DisableGateway

Constructors

DisableGateway

Fields

_dgjGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dgjGatewayARN :: Iso' DisableGateway Text

data DisableGatewayResponse

Constructors

DisableGatewayResponse

Fields

_dgpGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dgpGatewayARN :: Iso' DisableGatewayResponse (Maybe Text)