Network.AWS.StorageGateway.V2013_06_30.ListGateways

Description

This operation lists gateways owned by an AWS account in a region specified in the request. The returned list is ordered by gateway Amazon Resource Name (ARN). By default, the operation returns a maximum of 100 gateways. This operation supports pagination that allows you to optionally reduce the number of gateways returned in a response. If you have more gateways than are returned in a response-that is, the response returns only a truncated list of your gateways-the response contains a marker that you can specify in your next request to fetch the next page of gateways. List Gateways The following example does not specify any criteria for the returned list. Note that the request body is "{}". The response returns gateways (or up to the first 100) in the specified region owned by the AWS account. POST / HTTP/1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.ListGateways HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 178 { GatewayList: [ { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway", GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway2" } ] }.

Documentation

listGateways :: ListGateways

Minimum specification for a ListGateways request.

data ListGateways

Constructors

ListGateways

Fields

_lgiMarker :: Maybe Text

An opaque string that indicates the position at which to begin the returned list of gateways.

_lgiLimit :: Maybe Integer

Specifies that the list of gateways returned be limited to the specified number of items.

lgiMarker :: Lens' ListGateways (Maybe Text)

lgiLimit :: Lens' ListGateways (Maybe Integer)

data ListGatewaysResponse

Constructors

ListGatewaysResponse

Fields

_lgoGateways :: [GatewayInformation]

_lgoMarker :: Maybe Text

lgoMarker :: Lens' ListGatewaysResponse (Maybe Text)

lgoGateways :: Lens' ListGatewaysResponse [GatewayInformation]