Network.AWS.StorageGateway.V2013_06_30.ListVolumeRecoveryPoints

Description

This operation lists the recovery points for a specified gateway. This operation is supported only for the gateway-cached volume architecture. Each gateway-cached volume has one recovery point. A volume recovery point is a point in time at which all data of the volume is consistent and from which you can create a snapshot. To create a snapshot from a volume recovery point use the CreateSnapshotFromVolumeRecoveryPoint operation. Example Request The following example sends a ListVolumeRecoveryPoints request to take a snapshot of the specified example volume. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com Content-Type: application/x-amz-json-1.1 Authorization: AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE20120425us-east-1storagegatewayaws4_request, SignedHeaders=content-type;host;x-amz-date;x-amz-target, Signature=9cd5a3584d1d67d57e61f120f35102d6b3649066abdd4bf4bbcf05bd9f2f8fe2 x-amz-date: 20120912T120000Z x-amz-target: StorageGateway_20120630.ListVolumeRecoveryPoints { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" } HTTP/1.1 200 OK x-amzn-RequestId: gur28r2rqlgb8vvs0mq17hlgij1q8glle1qeu3kpgg6f0kstauu0 Date: Wed, 12 Sep 2012 12:00:02 GMT Content-Type: application/x-amz-json-1.1 Content-length: 137 { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway", VolumeRecoveryPointInfos: [ { VolumeARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewayvolume/vol-1122AABB", VolumeRecoveryPointTime: "2012-09-04T21:08:44.627Z", VolumeSizeInBytes: 536870912000, VolumeUsageInBytes: 6694048 } ] }.

Documentation

data ListVolumeRecoveryPoints

Constructors

ListVolumeRecoveryPoints

Fields

_lvrpiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

lvrpiGatewayARN :: Iso' ListVolumeRecoveryPoints Text

data ListVolumeRecoveryPointsResponse

Constructors

ListVolumeRecoveryPointsResponse

Fields

_lvrpoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_lvrpoVolumeRecoveryPointInfos :: [VolumeRecoveryPointInfo]

lvrpoVolumeRecoveryPointInfos :: Lens' ListVolumeRecoveryPointsResponse [VolumeRecoveryPointInfo]

lvrpoGatewayARN :: Lens' ListVolumeRecoveryPointsResponse (Maybe Text)