Network.AWS.StorageGateway.V2013_06_30.ListVolumes

Description

This operation lists the iSCSI stored volumes of a gateway. Results are sorted by volume ARN. The response includes only the volume ARNs. If you want additional volume information, use the DescribeStorediSCSIVolumes API. The operation supports pagination. By default, the operation returns a maximum of up to 100 volumes. You can optionally specify the Limit field in the body to limit the number of volumes in the response. If the number of volumes returned in the response is truncated, the response includes a Marker field. You can use this Marker value in your subsequent request to retrieve the next set of volumes. Example Request The List iSCSI Volumes request in this example does not specify a limit or marker field in the response body. The response returns the volumes (up to the first 100) of the gateway. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.ListVolumes { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 346 { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway", VolumeInfos: [ { VolumeARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewayvolume/vol-1122AABB", VolumeType: STORED }, { VolumeARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewayvolume/vol-3344CCDD", VolumeType: STORED }, ] }.

Documentation

listVolumes

Arguments

:: Text
_lviGatewayARN

-> ListVolumes

Minimum specification for a ListVolumes request.

data ListVolumes

Constructors

ListVolumes

Fields

_lviGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_lviMarker :: Maybe Text

A string that indicates the position at which to begin the returned list of volumes. Obtain the marker from the response of a previous List iSCSI Volumes request.

_lviLimit :: Maybe Integer

Specifies that the list of volumes returned be limited to the specified number of items.

lviMarker :: Lens' ListVolumes (Maybe Text)

lviLimit :: Lens' ListVolumes (Maybe Integer)

lviGatewayARN :: Lens' ListVolumes Text

data ListVolumesResponse

Constructors

ListVolumesResponse

Fields

_lvoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_lvoMarker :: Maybe Text

_lvoVolumeInfos :: [VolumeInformation]

lvoVolumeInfos :: Lens' ListVolumesResponse [VolumeInformation]

lvoMarker :: Lens' ListVolumesResponse (Maybe Text)

lvoGatewayARN :: Lens' ListVolumesResponse (Maybe Text)