Network.AWS.StorageGateway.V2013_06_30.UpdateBandwidthRateLimit

Description

This operation updates the bandwidth rate limits of a gateway. You can update both the upload and download bandwidth rate limit or specify only one of the two. If you don't set a bandwidth rate limit, the existing rate limit remains. By default, a gateway's bandwidth rate limits are not set. If you don't set any limit, the gateway does not have any limitations on its bandwidth usage and could potentially use the maximum available bandwidth. To specify which gateway to update, use the Amazon Resource Name (ARN) of the gateway in your request. Example Request The following example shows a request that returns the bandwidth throttle properties of a gateway. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.UpdateBandwidthRateLimit { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway", AverageUploadRateLimitInBitsPerSec: 51200, AverageDownloadRateLimitInBitsPerSec: 102400 } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 80 { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" }.

Documentation

updateBandwidthRateLimit

Arguments

:: Text
_ubrliGatewayARN

-> UpdateBandwidthRateLimit

Minimum specification for a UpdateBandwidthRateLimit request.

data UpdateBandwidthRateLimit

Constructors

UpdateBandwidthRateLimit

Fields

_ubrliGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_ubrliAverageDownloadRateLimitInBitsPerSec :: Maybe Integer

The average download bandwidth rate limit in bits per second.

_ubrliAverageUploadRateLimitInBitsPerSec :: Maybe Integer

The average upload bandwidth rate limit in bits per second.

ubrliGatewayARN :: Lens' UpdateBandwidthRateLimit Text

ubrliAverageUploadRateLimitInBitsPerSec :: Lens' UpdateBandwidthRateLimit (Maybe Integer)

ubrliAverageDownloadRateLimitInBitsPerSec :: Lens' UpdateBandwidthRateLimit (Maybe Integer)

data UpdateBandwidthRateLimitResponse

Constructors

UpdateBandwidthRateLimitResponse

Fields

_ubrloGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

ubrloGatewayARN :: Iso' UpdateBandwidthRateLimitResponse (Maybe Text)