Network.AWS.Support.AddCommunicationToCase

Description

Adds additional customer communication to an AWS Support case. Use the caseId parameter to identify the case to which to add communication. You can list a set of email addresses to copy on the communication by using the ccEmailAddresses parameter. The communicationBody value contains the text of the communication.

You must have a Business or Enterprise Support plan to use the AWS Support API.

If you call the AWS Support API from an account that does not have a Business or Enterprise Support plan, the SubscriptionRequiredException error message appears. For information about changing your support plan, see AWS Support.

Creating a Request

data AddCommunicationToCase

See: newAddCommunicationToCase smart constructor.

Constructors

AddCommunicationToCase'

Fields

caseId :: Maybe Text

The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

ccEmailAddresses :: Maybe [Text]

The email addresses in the CC line of an email to be added to the support case.

attachmentSetId :: Maybe Text

The ID of a set of one or more attachments for the communication to add to the case. Create the set by calling AddAttachmentsToSet

communicationBody :: Text

The body of an email communication to add to the support case.

newAddCommunicationToCase

Arguments

:: Text
$sel:communicationBody:AddCommunicationToCase'

-> AddCommunicationToCase

Create a value of AddCommunicationToCase with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:caseId:AddCommunicationToCase', addCommunicationToCase_caseId - The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

$sel:ccEmailAddresses:AddCommunicationToCase', addCommunicationToCase_ccEmailAddresses - The email addresses in the CC line of an email to be added to the support case.

$sel:attachmentSetId:AddCommunicationToCase', addCommunicationToCase_attachmentSetId - The ID of a set of one or more attachments for the communication to add to the case. Create the set by calling AddAttachmentsToSet

$sel:communicationBody:AddCommunicationToCase', addCommunicationToCase_communicationBody - The body of an email communication to add to the support case.

Request Lenses

addCommunicationToCase_caseId :: Lens' AddCommunicationToCase (Maybe Text)

The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

addCommunicationToCase_ccEmailAddresses :: Lens' AddCommunicationToCase (Maybe [Text])

The email addresses in the CC line of an email to be added to the support case.

addCommunicationToCase_attachmentSetId :: Lens' AddCommunicationToCase (Maybe Text)

The ID of a set of one or more attachments for the communication to add to the case. Create the set by calling AddAttachmentsToSet

addCommunicationToCase_communicationBody :: Lens' AddCommunicationToCase Text

The body of an email communication to add to the support case.

Destructuring the Response

data AddCommunicationToCaseResponse

The result of the AddCommunicationToCase operation.

See: newAddCommunicationToCaseResponse smart constructor.

Constructors

AddCommunicationToCaseResponse'

Fields

result :: Maybe Bool

True if AddCommunicationToCase succeeds. Otherwise, returns an error.

httpStatus :: Int

The response's http status code.

newAddCommunicationToCaseResponse

Arguments

:: Int
$sel:httpStatus:AddCommunicationToCaseResponse'

-> AddCommunicationToCaseResponse

Create a value of AddCommunicationToCaseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:result:AddCommunicationToCaseResponse', addCommunicationToCaseResponse_result - True if AddCommunicationToCase succeeds. Otherwise, returns an error.

$sel:httpStatus:AddCommunicationToCaseResponse', addCommunicationToCaseResponse_httpStatus - The response's http status code.

Response Lenses

addCommunicationToCaseResponse_result :: Lens' AddCommunicationToCaseResponse (Maybe Bool)

True if AddCommunicationToCase succeeds. Otherwise, returns an error.

addCommunicationToCaseResponse_httpStatus :: Lens' AddCommunicationToCaseResponse Int

The response's http status code.