Network.AWS.Support.CreateCase

Description

Creates a case in the AWS Support Center. This operation is similar to how you create a case in the AWS Support Center Create Case page.

The AWS Support API doesn't support requesting service limit increases. You can submit a service limit increase in the following ways:

Submit a request from the AWS Support Center Create Case page.

Use the Service Quotas RequestServiceQuotaIncrease operation.

A successful CreateCase request returns an AWS Support case number. You can use the DescribeCases operation and specify the case number to get existing AWS Support cases. After you create a case, use the AddCommunicationToCase operation to add additional communication or attachments to an existing case.

The caseId is separate from the displayId that appears in the AWS Support Center. Use the DescribeCases operation to get the displayId.

You must have a Business or Enterprise Support plan to use the AWS Support API.

If you call the AWS Support API from an account that does not have a Business or Enterprise Support plan, the SubscriptionRequiredException error message appears. For information about changing your support plan, see AWS Support.

Creating a Request

data CreateCase

See: newCreateCase smart constructor.

Constructors

CreateCase'

Fields

serviceCode :: Maybe Text

The code for the AWS service. You can use the DescribeServices operation to get the possible serviceCode values.

categoryCode :: Maybe Text

The category of problem for the support case. You also use the DescribeServices operation to get the category code for a service. Each AWS service defines its own set of category codes.

ccEmailAddresses :: Maybe [Text]

A list of email addresses that AWS Support copies on case correspondence. AWS Support identifies the account that creates the case when you specify your AWS credentials in an HTTP POST method or use the AWS SDKs.

issueType :: Maybe Text

The type of issue for the case. You can specify customer-service or technical. If you don't specify a value, the default is technical.

attachmentSetId :: Maybe Text

The ID of a set of one or more attachments for the case. Create the set by using the AddAttachmentsToSet operation.

severityCode :: Maybe Text

A value that indicates the urgency of the case. This value determines the response time according to your service level agreement with AWS Support. You can use the DescribeSeverityLevels operation to get the possible values for severityCode.

For more information, see SeverityLevel and Choosing a Severity in the AWS Support User Guide.

The availability of severity levels depends on the support plan for the AWS account.

language :: Maybe Text

The language in which AWS Support handles the case. You must specify the ISO 639-1 code for the language parameter if you want support in that language. Currently, English ("en") and Japanese ("ja") are supported.

subject :: Text

The title of the support case. The title appears in the Subject field on the AWS Support Center Create Case page.

communicationBody :: Text

The communication body text that describes the issue. This text appears in the Description field on the AWS Support Center Create Case page.

newCreateCase

Arguments

:: Text
$sel:subject:CreateCase'

-> Text
$sel:communicationBody:CreateCase'

-> CreateCase

Create a value of CreateCase with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceCode:CreateCase', createCase_serviceCode - The code for the AWS service. You can use the DescribeServices operation to get the possible serviceCode values.

$sel:categoryCode:CreateCase', createCase_categoryCode - The category of problem for the support case. You also use the DescribeServices operation to get the category code for a service. Each AWS service defines its own set of category codes.

$sel:ccEmailAddresses:CreateCase', createCase_ccEmailAddresses - A list of email addresses that AWS Support copies on case correspondence. AWS Support identifies the account that creates the case when you specify your AWS credentials in an HTTP POST method or use the AWS SDKs.

$sel:issueType:CreateCase', createCase_issueType - The type of issue for the case. You can specify customer-service or technical. If you don't specify a value, the default is technical.

$sel:attachmentSetId:CreateCase', createCase_attachmentSetId - The ID of a set of one or more attachments for the case. Create the set by using the AddAttachmentsToSet operation.

$sel:severityCode:CreateCase', createCase_severityCode - A value that indicates the urgency of the case. This value determines the response time according to your service level agreement with AWS Support. You can use the DescribeSeverityLevels operation to get the possible values for severityCode.

For more information, see SeverityLevel and Choosing a Severity in the AWS Support User Guide.

The availability of severity levels depends on the support plan for the AWS account.

$sel:language:CreateCase', createCase_language - The language in which AWS Support handles the case. You must specify the ISO 639-1 code for the language parameter if you want support in that language. Currently, English ("en") and Japanese ("ja") are supported.

$sel:subject:CreateCase', createCase_subject - The title of the support case. The title appears in the Subject field on the AWS Support Center Create Case page.

$sel:communicationBody:CreateCase', createCase_communicationBody - The communication body text that describes the issue. This text appears in the Description field on the AWS Support Center Create Case page.

Request Lenses

createCase_serviceCode :: Lens' CreateCase (Maybe Text)

The code for the AWS service. You can use the DescribeServices operation to get the possible serviceCode values.

createCase_categoryCode :: Lens' CreateCase (Maybe Text)

The category of problem for the support case. You also use the DescribeServices operation to get the category code for a service. Each AWS service defines its own set of category codes.

createCase_ccEmailAddresses :: Lens' CreateCase (Maybe [Text])

A list of email addresses that AWS Support copies on case correspondence. AWS Support identifies the account that creates the case when you specify your AWS credentials in an HTTP POST method or use the AWS SDKs.

createCase_issueType :: Lens' CreateCase (Maybe Text)

The type of issue for the case. You can specify customer-service or technical. If you don't specify a value, the default is technical.

createCase_attachmentSetId :: Lens' CreateCase (Maybe Text)

The ID of a set of one or more attachments for the case. Create the set by using the AddAttachmentsToSet operation.

createCase_severityCode :: Lens' CreateCase (Maybe Text)

A value that indicates the urgency of the case. This value determines the response time according to your service level agreement with AWS Support. You can use the DescribeSeverityLevels operation to get the possible values for severityCode.

For more information, see SeverityLevel and Choosing a Severity in the AWS Support User Guide.

The availability of severity levels depends on the support plan for the AWS account.

createCase_language :: Lens' CreateCase (Maybe Text)

The language in which AWS Support handles the case. You must specify the ISO 639-1 code for the language parameter if you want support in that language. Currently, English ("en") and Japanese ("ja") are supported.

createCase_subject :: Lens' CreateCase Text

The title of the support case. The title appears in the Subject field on the AWS Support Center Create Case page.

createCase_communicationBody :: Lens' CreateCase Text

The communication body text that describes the issue. This text appears in the Description field on the AWS Support Center Create Case page.

Destructuring the Response

data CreateCaseResponse

The support case ID returned by a successful completion of the CreateCase operation.

See: newCreateCaseResponse smart constructor.

Constructors

CreateCaseResponse'

Fields

caseId :: Maybe Text

The support case ID requested or returned in the call. The case ID is an alphanumeric string in the following format: case-12345678910-2013-c4c1d2bf33c5cf47

httpStatus :: Int

The response's http status code.

newCreateCaseResponse

Arguments

:: Int
$sel:httpStatus:CreateCaseResponse'

-> CreateCaseResponse

Create a value of CreateCaseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:caseId:CreateCaseResponse', createCaseResponse_caseId - The support case ID requested or returned in the call. The case ID is an alphanumeric string in the following format: case-12345678910-2013-c4c1d2bf33c5cf47

$sel:httpStatus:CreateCaseResponse', createCaseResponse_httpStatus - The response's http status code.

Response Lenses

createCaseResponse_caseId :: Lens' CreateCaseResponse (Maybe Text)

The support case ID requested or returned in the call. The case ID is an alphanumeric string in the following format: case-12345678910-2013-c4c1d2bf33c5cf47

createCaseResponse_httpStatus :: Lens' CreateCaseResponse Int

The response's http status code.