Network.AWS.Support.DescribeServices

Description

Returns the current list of AWS services and a list of service categories for each service. You then use service names and categories in your CreateCase requests. Each AWS service has its own set of categories.

The service codes and category codes correspond to the values that appear in the Service and Category lists on the AWS Support Center Create Case page. The values in those fields don't necessarily match the service codes and categories returned by the DescribeServices operation. Always use the service codes and categories that the DescribeServices operation returns, so that you have the most recent set of service and category codes.

You must have a Business or Enterprise Support plan to use the AWS Support API.

If you call the AWS Support API from an account that does not have a Business or Enterprise Support plan, the SubscriptionRequiredException error message appears. For information about changing your support plan, see AWS Support.

Creating a Request

data DescribeServices

See: newDescribeServices smart constructor.

Constructors

DescribeServices'

Fields

serviceCodeList :: Maybe [Text]

A JSON-formatted list of service codes available for AWS services.

language :: Maybe Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

newDescribeServices :: DescribeServices

Create a value of DescribeServices with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceCodeList:DescribeServices', describeServices_serviceCodeList - A JSON-formatted list of service codes available for AWS services.

$sel:language:DescribeServices', describeServices_language - The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

Request Lenses

describeServices_serviceCodeList :: Lens' DescribeServices (Maybe [Text])

A JSON-formatted list of service codes available for AWS services.

describeServices_language :: Lens' DescribeServices (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

Destructuring the Response

data DescribeServicesResponse

The list of AWS services returned by the DescribeServices operation.

See: newDescribeServicesResponse smart constructor.

Constructors

DescribeServicesResponse'

Fields

services :: Maybe [SupportService]

A JSON-formatted list of AWS services.

httpStatus :: Int

The response's http status code.

newDescribeServicesResponse

Arguments

:: Int
$sel:httpStatus:DescribeServicesResponse'

-> DescribeServicesResponse

Create a value of DescribeServicesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:services:DescribeServicesResponse', describeServicesResponse_services - A JSON-formatted list of AWS services.

$sel:httpStatus:DescribeServicesResponse', describeServicesResponse_httpStatus - The response's http status code.

Response Lenses

describeServicesResponse_services :: Lens' DescribeServicesResponse (Maybe [SupportService])

A JSON-formatted list of AWS services.

describeServicesResponse_httpStatus :: Lens' DescribeServicesResponse Int

The response's http status code.