Network.AWS.Support.DescribeTrustedAdvisorCheckRefreshStatuses

Description

Returns the refresh status of the AWS Trusted Advisor checks that have the specified check IDs. You can get the check IDs by calling the DescribeTrustedAdvisorChecks operation.

Some checks are refreshed automatically, and you can't return their refresh statuses by using the DescribeTrustedAdvisorCheckRefreshStatuses operation. If you call this operation for these checks, you might see an InvalidParameterValue error.

You must have a Business or Enterprise Support plan to use the AWS Support API.

If you call the AWS Support API from an account that does not have a Business or Enterprise Support plan, the SubscriptionRequiredException error message appears. For information about changing your support plan, see AWS Support.

Creating a Request

data DescribeTrustedAdvisorCheckRefreshStatuses

See: newDescribeTrustedAdvisorCheckRefreshStatuses smart constructor.

Constructors

DescribeTrustedAdvisorCheckRefreshStatuses'

Fields

checkIds :: [Text]

The IDs of the Trusted Advisor checks to get the status.

If you specify the check ID of a check that is automatically refreshed, you might see an InvalidParameterValue error.

newDescribeTrustedAdvisorCheckRefreshStatuses :: DescribeTrustedAdvisorCheckRefreshStatuses

Create a value of DescribeTrustedAdvisorCheckRefreshStatuses with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checkIds:DescribeTrustedAdvisorCheckRefreshStatuses', describeTrustedAdvisorCheckRefreshStatuses_checkIds - The IDs of the Trusted Advisor checks to get the status.

If you specify the check ID of a check that is automatically refreshed, you might see an InvalidParameterValue error.

Request Lenses

describeTrustedAdvisorCheckRefreshStatuses_checkIds :: Lens' DescribeTrustedAdvisorCheckRefreshStatuses [Text]

The IDs of the Trusted Advisor checks to get the status.

If you specify the check ID of a check that is automatically refreshed, you might see an InvalidParameterValue error.

Destructuring the Response

data DescribeTrustedAdvisorCheckRefreshStatusesResponse

The statuses of the Trusted Advisor checks returned by the DescribeTrustedAdvisorCheckRefreshStatuses operation.

See: newDescribeTrustedAdvisorCheckRefreshStatusesResponse smart constructor.

Constructors

DescribeTrustedAdvisorCheckRefreshStatusesResponse'

Fields

httpStatus :: Int

The response's http status code.

statuses :: [TrustedAdvisorCheckRefreshStatus]

The refresh status of the specified Trusted Advisor checks.

newDescribeTrustedAdvisorCheckRefreshStatusesResponse

Arguments

:: Int
$sel:httpStatus:DescribeTrustedAdvisorCheckRefreshStatusesResponse'

-> DescribeTrustedAdvisorCheckRefreshStatusesResponse

Create a value of DescribeTrustedAdvisorCheckRefreshStatusesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeTrustedAdvisorCheckRefreshStatusesResponse', describeTrustedAdvisorCheckRefreshStatusesResponse_httpStatus - The response's http status code.

$sel:statuses:DescribeTrustedAdvisorCheckRefreshStatusesResponse', describeTrustedAdvisorCheckRefreshStatusesResponse_statuses - The refresh status of the specified Trusted Advisor checks.

Response Lenses

describeTrustedAdvisorCheckRefreshStatusesResponse_httpStatus :: Lens' DescribeTrustedAdvisorCheckRefreshStatusesResponse Int

The response's http status code.

describeTrustedAdvisorCheckRefreshStatusesResponse_statuses :: Lens' DescribeTrustedAdvisorCheckRefreshStatusesResponse [TrustedAdvisorCheckRefreshStatus]

The refresh status of the specified Trusted Advisor checks.