Network.AWS.Support.DescribeTrustedAdvisorChecks

Description

Returns information about all available AWS Trusted Advisor checks, including the name, ID, category, description, and metadata. You must specify a language code. The AWS Support API currently supports English ("en") and Japanese ("ja"). The response contains a TrustedAdvisorCheckDescription object for each check. You must set the AWS Region to us-east-1.

You must have a Business or Enterprise Support plan to use the AWS Support API.

If you call the AWS Support API from an account that does not have a Business or Enterprise Support plan, the SubscriptionRequiredException error message appears. For information about changing your support plan, see AWS Support.

The names and descriptions for Trusted Advisor checks are subject to change. We recommend that you specify the check ID in your code to uniquely identify a check.

Creating a Request

data DescribeTrustedAdvisorChecks

See: newDescribeTrustedAdvisorChecks smart constructor.

Constructors

DescribeTrustedAdvisorChecks'

Fields

language :: Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

newDescribeTrustedAdvisorChecks

Arguments

:: Text
$sel:language:DescribeTrustedAdvisorChecks'

-> DescribeTrustedAdvisorChecks

Create a value of DescribeTrustedAdvisorChecks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:language:DescribeTrustedAdvisorChecks', describeTrustedAdvisorChecks_language - The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

Request Lenses

describeTrustedAdvisorChecks_language :: Lens' DescribeTrustedAdvisorChecks Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

Destructuring the Response

data DescribeTrustedAdvisorChecksResponse

Information about the Trusted Advisor checks returned by the DescribeTrustedAdvisorChecks operation.

See: newDescribeTrustedAdvisorChecksResponse smart constructor.

Constructors

DescribeTrustedAdvisorChecksResponse'

Fields

httpStatus :: Int

The response's http status code.

checks :: [TrustedAdvisorCheckDescription]

Information about all available Trusted Advisor checks.

newDescribeTrustedAdvisorChecksResponse

Arguments

:: Int
$sel:httpStatus:DescribeTrustedAdvisorChecksResponse'

-> DescribeTrustedAdvisorChecksResponse

Create a value of DescribeTrustedAdvisorChecksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeTrustedAdvisorChecksResponse', describeTrustedAdvisorChecksResponse_httpStatus - The response's http status code.

$sel:checks:DescribeTrustedAdvisorChecksResponse', describeTrustedAdvisorChecksResponse_checks - Information about all available Trusted Advisor checks.

Response Lenses

describeTrustedAdvisorChecksResponse_httpStatus :: Lens' DescribeTrustedAdvisorChecksResponse Int

The response's http status code.

describeTrustedAdvisorChecksResponse_checks :: Lens' DescribeTrustedAdvisorChecksResponse [TrustedAdvisorCheckDescription]

Information about all available Trusted Advisor checks.