Network.AWS.Support.Lens

Description

Operations

RefreshTrustedAdvisorCheck

refreshTrustedAdvisorCheck_checkId :: Lens' RefreshTrustedAdvisorCheck Text

The unique identifier for the Trusted Advisor check to refresh.

Specifying the check ID of a check that is automatically refreshed causes an InvalidParameterValue error.

refreshTrustedAdvisorCheckResponse_httpStatus :: Lens' RefreshTrustedAdvisorCheckResponse Int

The response's http status code.

refreshTrustedAdvisorCheckResponse_status :: Lens' RefreshTrustedAdvisorCheckResponse TrustedAdvisorCheckRefreshStatus

The current refresh status for a check, including the amount of time until the check is eligible for refresh.

DescribeCases

describeCases_includeResolvedCases :: Lens' DescribeCases (Maybe Bool)

Specifies whether to include resolved support cases in the DescribeCases response. By default, resolved cases aren't included.

describeCases_caseIdList :: Lens' DescribeCases (Maybe [Text])

A list of ID numbers of the support cases you want returned. The maximum number of cases is 100.

describeCases_afterTime :: Lens' DescribeCases (Maybe Text)

The start date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

describeCases_beforeTime :: Lens' DescribeCases (Maybe Text)

The end date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

describeCases_nextToken :: Lens' DescribeCases (Maybe Text)

A resumption point for pagination.

describeCases_includeCommunications :: Lens' DescribeCases (Maybe Bool)

Specifies whether to include communications in the DescribeCases response. By default, communications are included.

describeCases_displayId :: Lens' DescribeCases (Maybe Text)

The ID displayed for a case in the AWS Support Center user interface.

describeCases_language :: Lens' DescribeCases (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

describeCases_maxResults :: Lens' DescribeCases (Maybe Natural)

The maximum number of results to return before paginating.

describeCasesResponse_cases :: Lens' DescribeCasesResponse (Maybe [CaseDetails])

The details for the cases that match the request.

describeCasesResponse_nextToken :: Lens' DescribeCasesResponse (Maybe Text)

A resumption point for pagination.

describeCasesResponse_httpStatus :: Lens' DescribeCasesResponse Int

The response's http status code.

DescribeTrustedAdvisorCheckRefreshStatuses

describeTrustedAdvisorCheckRefreshStatuses_checkIds :: Lens' DescribeTrustedAdvisorCheckRefreshStatuses [Text]

The IDs of the Trusted Advisor checks to get the status.

If you specify the check ID of a check that is automatically refreshed, you might see an InvalidParameterValue error.

describeTrustedAdvisorCheckRefreshStatusesResponse_httpStatus :: Lens' DescribeTrustedAdvisorCheckRefreshStatusesResponse Int

The response's http status code.

describeTrustedAdvisorCheckRefreshStatusesResponse_statuses :: Lens' DescribeTrustedAdvisorCheckRefreshStatusesResponse [TrustedAdvisorCheckRefreshStatus]

The refresh status of the specified Trusted Advisor checks.

DescribeTrustedAdvisorCheckSummaries

describeTrustedAdvisorCheckSummaries_checkIds :: Lens' DescribeTrustedAdvisorCheckSummaries [Text]

The IDs of the Trusted Advisor checks.

describeTrustedAdvisorCheckSummariesResponse_httpStatus :: Lens' DescribeTrustedAdvisorCheckSummariesResponse Int

The response's http status code.

describeTrustedAdvisorCheckSummariesResponse_summaries :: Lens' DescribeTrustedAdvisorCheckSummariesResponse [TrustedAdvisorCheckSummary]

The summary information for the requested Trusted Advisor checks.

CreateCase

createCase_severityCode :: Lens' CreateCase (Maybe Text)

A value that indicates the urgency of the case. This value determines the response time according to your service level agreement with AWS Support. You can use the DescribeSeverityLevels operation to get the possible values for severityCode.

For more information, see SeverityLevel and Choosing a Severity in the AWS Support User Guide.

The availability of severity levels depends on the support plan for the AWS account.

createCase_issueType :: Lens' CreateCase (Maybe Text)

The type of issue for the case. You can specify customer-service or technical. If you don't specify a value, the default is technical.

createCase_ccEmailAddresses :: Lens' CreateCase (Maybe [Text])

A list of email addresses that AWS Support copies on case correspondence. AWS Support identifies the account that creates the case when you specify your AWS credentials in an HTTP POST method or use the AWS SDKs.

createCase_language :: Lens' CreateCase (Maybe Text)

The language in which AWS Support handles the case. You must specify the ISO 639-1 code for the language parameter if you want support in that language. Currently, English ("en") and Japanese ("ja") are supported.

createCase_categoryCode :: Lens' CreateCase (Maybe Text)

The category of problem for the support case. You also use the DescribeServices operation to get the category code for a service. Each AWS service defines its own set of category codes.

createCase_serviceCode :: Lens' CreateCase (Maybe Text)

The code for the AWS service. You can use the DescribeServices operation to get the possible serviceCode values.

createCase_attachmentSetId :: Lens' CreateCase (Maybe Text)

The ID of a set of one or more attachments for the case. Create the set by using the AddAttachmentsToSet operation.

createCase_subject :: Lens' CreateCase Text

The title of the support case. The title appears in the Subject field on the AWS Support Center Create Case page.

createCase_communicationBody :: Lens' CreateCase Text

The communication body text that describes the issue. This text appears in the Description field on the AWS Support Center Create Case page.

createCaseResponse_caseId :: Lens' CreateCaseResponse (Maybe Text)

The support case ID requested or returned in the call. The case ID is an alphanumeric string in the following format: case-12345678910-2013-c4c1d2bf33c5cf47

createCaseResponse_httpStatus :: Lens' CreateCaseResponse Int

The response's http status code.

ResolveCase

resolveCase_caseId :: Lens' ResolveCase (Maybe Text)

The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

resolveCaseResponse_initialCaseStatus :: Lens' ResolveCaseResponse (Maybe Text)

The status of the case when the ResolveCase request was sent.

resolveCaseResponse_finalCaseStatus :: Lens' ResolveCaseResponse (Maybe Text)

The status of the case after the ResolveCase request was processed.

resolveCaseResponse_httpStatus :: Lens' ResolveCaseResponse Int

The response's http status code.

DescribeSeverityLevels

describeSeverityLevels_language :: Lens' DescribeSeverityLevels (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

describeSeverityLevelsResponse_severityLevels :: Lens' DescribeSeverityLevelsResponse (Maybe [SeverityLevel])

The available severity levels for the support case. Available severity levels are defined by your service level agreement with AWS.

describeSeverityLevelsResponse_httpStatus :: Lens' DescribeSeverityLevelsResponse Int

The response's http status code.

DescribeTrustedAdvisorChecks

describeTrustedAdvisorChecks_language :: Lens' DescribeTrustedAdvisorChecks Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

describeTrustedAdvisorChecksResponse_httpStatus :: Lens' DescribeTrustedAdvisorChecksResponse Int

The response's http status code.

describeTrustedAdvisorChecksResponse_checks :: Lens' DescribeTrustedAdvisorChecksResponse [TrustedAdvisorCheckDescription]

Information about all available Trusted Advisor checks.

DescribeAttachment

describeAttachment_attachmentId :: Lens' DescribeAttachment Text

The ID of the attachment to return. Attachment IDs are returned by the DescribeCommunications operation.

describeAttachmentResponse_attachment :: Lens' DescribeAttachmentResponse (Maybe Attachment)

This object includes the attachment content and file name.

In the previous response syntax, the value for the data parameter appears as blob, which is represented as a base64-encoded string. The value for fileName is the name of the attachment, such as troubleshoot-screenshot.png.

describeAttachmentResponse_httpStatus :: Lens' DescribeAttachmentResponse Int

The response's http status code.

AddAttachmentsToSet

addAttachmentsToSet_attachmentSetId :: Lens' AddAttachmentsToSet (Maybe Text)

The ID of the attachment set. If an attachmentSetId is not specified, a new attachment set is created, and the ID of the set is returned in the response. If an attachmentSetId is specified, the attachments are added to the specified set, if it exists.

addAttachmentsToSet_attachments :: Lens' AddAttachmentsToSet [Attachment]

One or more attachments to add to the set. You can add up to three attachments per set. The size limit is 5 MB per attachment.

In the Attachment object, use the data parameter to specify the contents of the attachment file. In the previous request syntax, the value for data appear as blob, which is represented as a base64-encoded string. The value for fileName is the name of the attachment, such as troubleshoot-screenshot.png.

addAttachmentsToSetResponse_expiryTime :: Lens' AddAttachmentsToSetResponse (Maybe Text)

The time and date when the attachment set expires.

addAttachmentsToSetResponse_attachmentSetId :: Lens' AddAttachmentsToSetResponse (Maybe Text)

The ID of the attachment set. If an attachmentSetId was not specified, a new attachment set is created, and the ID of the set is returned in the response. If an attachmentSetId was specified, the attachments are added to the specified set, if it exists.

addAttachmentsToSetResponse_httpStatus :: Lens' AddAttachmentsToSetResponse Int

The response's http status code.

DescribeTrustedAdvisorCheckResult

describeTrustedAdvisorCheckResult_language :: Lens' DescribeTrustedAdvisorCheckResult (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

describeTrustedAdvisorCheckResult_checkId :: Lens' DescribeTrustedAdvisorCheckResult Text

The unique identifier for the Trusted Advisor check.

describeTrustedAdvisorCheckResultResponse_result :: Lens' DescribeTrustedAdvisorCheckResultResponse (Maybe TrustedAdvisorCheckResult)

The detailed results of the Trusted Advisor check.

describeTrustedAdvisorCheckResultResponse_httpStatus :: Lens' DescribeTrustedAdvisorCheckResultResponse Int

The response's http status code.

DescribeServices

describeServices_serviceCodeList :: Lens' DescribeServices (Maybe [Text])

A JSON-formatted list of service codes available for AWS services.

describeServices_language :: Lens' DescribeServices (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

describeServicesResponse_services :: Lens' DescribeServicesResponse (Maybe [SupportService])

A JSON-formatted list of AWS services.

describeServicesResponse_httpStatus :: Lens' DescribeServicesResponse Int

The response's http status code.

DescribeCommunications

describeCommunications_afterTime :: Lens' DescribeCommunications (Maybe Text)

The start date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

describeCommunications_beforeTime :: Lens' DescribeCommunications (Maybe Text)

The end date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

describeCommunications_nextToken :: Lens' DescribeCommunications (Maybe Text)

A resumption point for pagination.

describeCommunications_maxResults :: Lens' DescribeCommunications (Maybe Natural)

The maximum number of results to return before paginating.

describeCommunications_caseId :: Lens' DescribeCommunications Text

The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

describeCommunicationsResponse_nextToken :: Lens' DescribeCommunicationsResponse (Maybe Text)

A resumption point for pagination.

describeCommunicationsResponse_communications :: Lens' DescribeCommunicationsResponse (Maybe [Communication])

The communications for the case.

describeCommunicationsResponse_httpStatus :: Lens' DescribeCommunicationsResponse Int

The response's http status code.

AddCommunicationToCase

addCommunicationToCase_caseId :: Lens' AddCommunicationToCase (Maybe Text)

The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

addCommunicationToCase_ccEmailAddresses :: Lens' AddCommunicationToCase (Maybe [Text])

The email addresses in the CC line of an email to be added to the support case.

addCommunicationToCase_attachmentSetId :: Lens' AddCommunicationToCase (Maybe Text)

The ID of a set of one or more attachments for the communication to add to the case. Create the set by calling AddAttachmentsToSet

addCommunicationToCase_communicationBody :: Lens' AddCommunicationToCase Text

The body of an email communication to add to the support case.

addCommunicationToCaseResponse_result :: Lens' AddCommunicationToCaseResponse (Maybe Bool)

True if AddCommunicationToCase succeeds. Otherwise, returns an error.

addCommunicationToCaseResponse_httpStatus :: Lens' AddCommunicationToCaseResponse Int

The response's http status code.

Types

Attachment

attachment_data :: Lens' Attachment (Maybe ByteString)

The content of the attachment file.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

attachment_fileName :: Lens' Attachment (Maybe Text)

The name of the attachment file.

AttachmentDetails

attachmentDetails_attachmentId :: Lens' AttachmentDetails (Maybe Text)

The ID of the attachment.

attachmentDetails_fileName :: Lens' AttachmentDetails (Maybe Text)

The file name of the attachment.

CaseDetails

caseDetails_subject :: Lens' CaseDetails (Maybe Text)

The subject line for the case in the AWS Support Center.

caseDetails_status :: Lens' CaseDetails (Maybe Text)

The status of the case.

Valid values:

opened

pending-customer-action

reopened

resolved

unassigned

work-in-progress

caseDetails_recentCommunications :: Lens' CaseDetails (Maybe RecentCaseCommunications)

The five most recent communications between you and AWS Support Center, including the IDs of any attachments to the communications. Also includes a nextToken that you can use to retrieve earlier communications.

caseDetails_severityCode :: Lens' CaseDetails (Maybe Text)

The code for the severity level returned by the call to DescribeSeverityLevels.

caseDetails_caseId :: Lens' CaseDetails (Maybe Text)

The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

caseDetails_ccEmailAddresses :: Lens' CaseDetails (Maybe [Text])

The email addresses that receive copies of communication about the case.

caseDetails_displayId :: Lens' CaseDetails (Maybe Text)

The ID displayed for the case in the AWS Support Center. This is a numeric string.

caseDetails_submittedBy :: Lens' CaseDetails (Maybe Text)

The email address of the account that submitted the case.

caseDetails_language :: Lens' CaseDetails (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

caseDetails_timeCreated :: Lens' CaseDetails (Maybe Text)

The time that the case was created in the AWS Support Center.

caseDetails_categoryCode :: Lens' CaseDetails (Maybe Text)

The category of problem for the support case.

caseDetails_serviceCode :: Lens' CaseDetails (Maybe Text)

The code for the AWS service. You can get a list of codes and the corresponding service names by calling DescribeServices.

Category

category_name :: Lens' Category (Maybe Text)

The category name for the support case.

category_code :: Lens' Category (Maybe Text)

The category code for the support case.

Communication

communication_body :: Lens' Communication (Maybe Text)

The text of the communication between the customer and AWS Support.

communication_caseId :: Lens' Communication (Maybe Text)

The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

communication_submittedBy :: Lens' Communication (Maybe Text)

The identity of the account that submitted, or responded to, the support case. Customer entries include the role or IAM user as well as the email address. For example, "AdminRole (Role) <janedoe@example.com>. Entries from the AWS Support team display "Amazon Web Services," and don't show an email address.

communication_timeCreated :: Lens' Communication (Maybe Text)

The time the communication was created.

communication_attachmentSet :: Lens' Communication (Maybe [AttachmentDetails])

Information about the attachments to the case communication.

RecentCaseCommunications

recentCaseCommunications_nextToken :: Lens' RecentCaseCommunications (Maybe Text)

A resumption point for pagination.

recentCaseCommunications_communications :: Lens' RecentCaseCommunications (Maybe [Communication])

The five most recent communications associated with the case.

SeverityLevel

severityLevel_name :: Lens' SeverityLevel (Maybe Text)

The name of the severity level that corresponds to the severity level code.

The values returned by the API are different from the values that appear in the AWS Support Center. For example, the API uses the code low, but the name appears as General guidance in Support Center.

The following are the API code names and how they appear in the console:

low - General guidance

normal - System impaired

high - Production system impaired

urgent - Production system down

critical - Business-critical system down

For more information, see Choosing a severity in the AWS Support User Guide.

severityLevel_code :: Lens' SeverityLevel (Maybe Text)

The code for case severity level.

Valid values: low | normal | high | urgent | critical

SupportService

supportService_categories :: Lens' SupportService (Maybe [Category])

A list of categories that describe the type of support issue a case describes. Categories consist of a category name and a category code. Category names and codes are passed to AWS Support when you call CreateCase.

supportService_name :: Lens' SupportService (Maybe Text)

The friendly name for an AWS service. The code element contains the corresponding code.

supportService_code :: Lens' SupportService (Maybe Text)

The code for an AWS service returned by the DescribeServices response. The name element contains the corresponding friendly name.

TrustedAdvisorCategorySpecificSummary

trustedAdvisorCategorySpecificSummary_costOptimizing :: Lens' TrustedAdvisorCategorySpecificSummary (Maybe TrustedAdvisorCostOptimizingSummary)

The summary information about cost savings for a Trusted Advisor check that is in the Cost Optimizing category.

TrustedAdvisorCheckDescription

trustedAdvisorCheckDescription_id :: Lens' TrustedAdvisorCheckDescription Text

The unique identifier for the Trusted Advisor check.

trustedAdvisorCheckDescription_name :: Lens' TrustedAdvisorCheckDescription Text

The display name for the Trusted Advisor check.

trustedAdvisorCheckDescription_description :: Lens' TrustedAdvisorCheckDescription Text

The description of the Trusted Advisor check, which includes the alert criteria and recommended operations (contains HTML markup).

trustedAdvisorCheckDescription_category :: Lens' TrustedAdvisorCheckDescription Text

The category of the Trusted Advisor check.

trustedAdvisorCheckDescription_metadata :: Lens' TrustedAdvisorCheckDescription [Text]

The column headings for the data returned by the Trusted Advisor check. The order of the headings corresponds to the order of the data in the Metadata element of the TrustedAdvisorResourceDetail for the check. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

TrustedAdvisorCheckRefreshStatus

trustedAdvisorCheckRefreshStatus_checkId :: Lens' TrustedAdvisorCheckRefreshStatus Text

The unique identifier for the Trusted Advisor check.

trustedAdvisorCheckRefreshStatus_status :: Lens' TrustedAdvisorCheckRefreshStatus Text

The status of the Trusted Advisor check for which a refresh has been requested:

none - The check is not refreshed or the non-success status exceeds the timeout

enqueued - The check refresh requests has entered the refresh queue

processing - The check refresh request is picked up by the rule processing engine

success - The check is successfully refreshed

abandoned - The check refresh has failed

trustedAdvisorCheckRefreshStatus_millisUntilNextRefreshable :: Lens' TrustedAdvisorCheckRefreshStatus Integer

The amount of time, in milliseconds, until the Trusted Advisor check is eligible for refresh.

TrustedAdvisorCheckResult

trustedAdvisorCheckResult_checkId :: Lens' TrustedAdvisorCheckResult Text

The unique identifier for the Trusted Advisor check.

trustedAdvisorCheckResult_timestamp :: Lens' TrustedAdvisorCheckResult Text

The time of the last refresh of the check.

trustedAdvisorCheckResult_status :: Lens' TrustedAdvisorCheckResult Text

The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

trustedAdvisorCheckResult_resourcesSummary :: Lens' TrustedAdvisorCheckResult TrustedAdvisorResourcesSummary

Undocumented member.

trustedAdvisorCheckResult_categorySpecificSummary :: Lens' TrustedAdvisorCheckResult TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

trustedAdvisorCheckResult_flaggedResources :: Lens' TrustedAdvisorCheckResult [TrustedAdvisorResourceDetail]

The details about each resource listed in the check result.

TrustedAdvisorCheckSummary

trustedAdvisorCheckSummary_hasFlaggedResources :: Lens' TrustedAdvisorCheckSummary (Maybe Bool)

Specifies whether the Trusted Advisor check has flagged resources.

trustedAdvisorCheckSummary_checkId :: Lens' TrustedAdvisorCheckSummary Text

The unique identifier for the Trusted Advisor check.

trustedAdvisorCheckSummary_timestamp :: Lens' TrustedAdvisorCheckSummary Text

The time of the last refresh of the check.

trustedAdvisorCheckSummary_status :: Lens' TrustedAdvisorCheckSummary Text

The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

trustedAdvisorCheckSummary_resourcesSummary :: Lens' TrustedAdvisorCheckSummary TrustedAdvisorResourcesSummary

Undocumented member.

trustedAdvisorCheckSummary_categorySpecificSummary :: Lens' TrustedAdvisorCheckSummary TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

TrustedAdvisorCostOptimizingSummary

trustedAdvisorCostOptimizingSummary_estimatedMonthlySavings :: Lens' TrustedAdvisorCostOptimizingSummary Double

The estimated monthly savings that might be realized if the recommended operations are taken.

trustedAdvisorCostOptimizingSummary_estimatedPercentMonthlySavings :: Lens' TrustedAdvisorCostOptimizingSummary Double

The estimated percentage of savings that might be realized if the recommended operations are taken.

TrustedAdvisorResourceDetail

trustedAdvisorResourceDetail_isSuppressed :: Lens' TrustedAdvisorResourceDetail (Maybe Bool)

Specifies whether the AWS resource was ignored by Trusted Advisor because it was marked as suppressed by the user.

trustedAdvisorResourceDetail_region :: Lens' TrustedAdvisorResourceDetail (Maybe Text)

The AWS Region in which the identified resource is located.

trustedAdvisorResourceDetail_status :: Lens' TrustedAdvisorResourceDetail Text

The status code for the resource identified in the Trusted Advisor check.

trustedAdvisorResourceDetail_resourceId :: Lens' TrustedAdvisorResourceDetail Text

The unique identifier for the identified resource.

trustedAdvisorResourceDetail_metadata :: Lens' TrustedAdvisorResourceDetail [Text]

Additional information about the identified resource. The exact metadata and its order can be obtained by inspecting the TrustedAdvisorCheckDescription object returned by the call to DescribeTrustedAdvisorChecks. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

TrustedAdvisorResourcesSummary

trustedAdvisorResourcesSummary_resourcesProcessed :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources that were analyzed by the Trusted Advisor check.

trustedAdvisorResourcesSummary_resourcesFlagged :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources that were flagged (listed) by the Trusted Advisor check.

trustedAdvisorResourcesSummary_resourcesIgnored :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources ignored by Trusted Advisor because information was unavailable.

trustedAdvisorResourcesSummary_resourcesSuppressed :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources ignored by Trusted Advisor because they were marked as suppressed by the user.