Network.AWS.Support.RefreshTrustedAdvisorCheck

Description

Refreshes the AWS Trusted Advisor check that you specify using the check ID. You can get the check IDs by calling the DescribeTrustedAdvisorChecks operation.

Some checks are refreshed automatically. If you call the RefreshTrustedAdvisorCheck operation to refresh them, you might see the InvalidParameterValue error.

The response contains a TrustedAdvisorCheckRefreshStatus object.

You must have a Business or Enterprise Support plan to use the AWS Support API.

If you call the AWS Support API from an account that does not have a Business or Enterprise Support plan, the SubscriptionRequiredException error message appears. For information about changing your support plan, see AWS Support.

Creating a Request

data RefreshTrustedAdvisorCheck

See: newRefreshTrustedAdvisorCheck smart constructor.

Constructors

RefreshTrustedAdvisorCheck'

Fields

checkId :: Text

The unique identifier for the Trusted Advisor check to refresh.

Specifying the check ID of a check that is automatically refreshed causes an InvalidParameterValue error.

newRefreshTrustedAdvisorCheck

Arguments

:: Text
$sel:checkId:RefreshTrustedAdvisorCheck'

-> RefreshTrustedAdvisorCheck

Create a value of RefreshTrustedAdvisorCheck with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checkId:RefreshTrustedAdvisorCheck', refreshTrustedAdvisorCheck_checkId - The unique identifier for the Trusted Advisor check to refresh.

Specifying the check ID of a check that is automatically refreshed causes an InvalidParameterValue error.

Request Lenses

refreshTrustedAdvisorCheck_checkId :: Lens' RefreshTrustedAdvisorCheck Text

The unique identifier for the Trusted Advisor check to refresh.

Specifying the check ID of a check that is automatically refreshed causes an InvalidParameterValue error.

Destructuring the Response

data RefreshTrustedAdvisorCheckResponse

The current refresh status of a Trusted Advisor check.

See: newRefreshTrustedAdvisorCheckResponse smart constructor.

Constructors

RefreshTrustedAdvisorCheckResponse'

Fields

httpStatus :: Int

The response's http status code.

status :: TrustedAdvisorCheckRefreshStatus

The current refresh status for a check, including the amount of time until the check is eligible for refresh.

newRefreshTrustedAdvisorCheckResponse

Arguments

:: Int
$sel:httpStatus:RefreshTrustedAdvisorCheckResponse'

-> TrustedAdvisorCheckRefreshStatus
$sel:status:RefreshTrustedAdvisorCheckResponse'

-> RefreshTrustedAdvisorCheckResponse

Create a value of RefreshTrustedAdvisorCheckResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RefreshTrustedAdvisorCheckResponse', refreshTrustedAdvisorCheckResponse_httpStatus - The response's http status code.

$sel:status:RefreshTrustedAdvisorCheckResponse', refreshTrustedAdvisorCheckResponse_status - The current refresh status for a check, including the amount of time until the check is eligible for refresh.

Response Lenses

refreshTrustedAdvisorCheckResponse_httpStatus :: Lens' RefreshTrustedAdvisorCheckResponse Int

The response's http status code.

refreshTrustedAdvisorCheckResponse_status :: Lens' RefreshTrustedAdvisorCheckResponse TrustedAdvisorCheckRefreshStatus

The current refresh status for a check, including the amount of time until the check is eligible for refresh.