Network.AWS.Support.Types

Description

Service Configuration

defaultService :: Service

API version 2013-04-15 of the Amazon Support SDK configuration.

Errors

_AttachmentSetExpired :: AsError a => Getting (First ServiceError) a ServiceError

The expiration time of the attachment set has passed. The set expires 1 hour after it is created.

_CaseCreationLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

The case creation limit for the account has been exceeded.

_AttachmentSetSizeLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

A limit for the size of an attachment set has been exceeded. The limits are three attachments and 5 MB per attachment.

_AttachmentSetIdNotFound :: AsError a => Getting (First ServiceError) a ServiceError

An attachment set with the specified ID could not be found.

_CaseIdNotFound :: AsError a => Getting (First ServiceError) a ServiceError

The requested caseId couldn't be located.

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

An internal server error occurred.

_AttachmentIdNotFound :: AsError a => Getting (First ServiceError) a ServiceError

An attachment with the specified ID could not be found.

_AttachmentLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

The limit for the number of attachment sets created in a short period of time has been exceeded.

_DescribeAttachmentLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

The limit for the number of DescribeAttachment requests in a short period of time has been exceeded.

Attachment

data Attachment

An attachment to a case communication. The attachment consists of the file name and the content of the file.

See: newAttachment smart constructor.

Constructors

Attachment'

Fields

data' :: Maybe Base64

The content of the attachment file.

fileName :: Maybe Text

The name of the attachment file.

newAttachment :: Attachment

Create a value of Attachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:data':Attachment', attachment_data - The content of the attachment file.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:fileName:Attachment', attachment_fileName - The name of the attachment file.

attachment_data :: Lens' Attachment (Maybe ByteString)

The content of the attachment file.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

attachment_fileName :: Lens' Attachment (Maybe Text)

The name of the attachment file.

AttachmentDetails

data AttachmentDetails

The file name and ID of an attachment to a case communication. You can use the ID to retrieve the attachment with the DescribeAttachment operation.

See: newAttachmentDetails smart constructor.

Constructors

AttachmentDetails'

Fields

attachmentId :: Maybe Text

The ID of the attachment.

fileName :: Maybe Text

The file name of the attachment.

newAttachmentDetails :: AttachmentDetails

Create a value of AttachmentDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attachmentId:AttachmentDetails', attachmentDetails_attachmentId - The ID of the attachment.

$sel:fileName:AttachmentDetails', attachmentDetails_fileName - The file name of the attachment.

attachmentDetails_attachmentId :: Lens' AttachmentDetails (Maybe Text)

The ID of the attachment.

attachmentDetails_fileName :: Lens' AttachmentDetails (Maybe Text)

The file name of the attachment.

CaseDetails

data CaseDetails

A JSON-formatted object that contains the metadata for a support case. It is contained in the response from a DescribeCases request. CaseDetails contains the following fields:

caseId - The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47.

categoryCode - The category of problem for the support case. Corresponds to the CategoryCode values returned by a call to DescribeServices.

displayId - The identifier for the case on pages in the AWS Support Center.

language - The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

nextToken - A resumption point for pagination.

recentCommunications - One or more Communication objects. Fields of these objects are attachments, body, caseId, submittedBy, and timeCreated.

serviceCode - The identifier for the AWS service that corresponds to the service code defined in the call to DescribeServices.

severityCode - The severity code assigned to the case. Contains one of the values returned by the call to DescribeSeverityLevels. The possible values are: low, normal, high, urgent, and critical.

status - The status of the case in the AWS Support Center. Valid values:

opened

pending-customer-action

reopened

resolved

unassigned

work-in-progress

subject - The subject line of the case.

submittedBy - The email address of the account that submitted the case.

timeCreated - The time the case was created, in ISO-8601 format.

See: newCaseDetails smart constructor.

Constructors

CaseDetails'

Fields

displayId :: Maybe Text

The ID displayed for the case in the AWS Support Center. This is a numeric string.

status :: Maybe Text

The status of the case.

Valid values:

opened

pending-customer-action

reopened

resolved

unassigned

work-in-progress

caseId :: Maybe Text

The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

recentCommunications :: Maybe RecentCaseCommunications

The five most recent communications between you and AWS Support Center, including the IDs of any attachments to the communications. Also includes a nextToken that you can use to retrieve earlier communications.

serviceCode :: Maybe Text

The code for the AWS service. You can get a list of codes and the corresponding service names by calling DescribeServices.

categoryCode :: Maybe Text

The category of problem for the support case.

submittedBy :: Maybe Text

The email address of the account that submitted the case.

subject :: Maybe Text

The subject line for the case in the AWS Support Center.

ccEmailAddresses :: Maybe [Text]

The email addresses that receive copies of communication about the case.

severityCode :: Maybe Text

The code for the severity level returned by the call to DescribeSeverityLevels.

timeCreated :: Maybe Text

The time that the case was created in the AWS Support Center.

language :: Maybe Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

newCaseDetails :: CaseDetails

Create a value of CaseDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:displayId:CaseDetails', caseDetails_displayId - The ID displayed for the case in the AWS Support Center. This is a numeric string.

$sel:status:CaseDetails', caseDetails_status - The status of the case.

Valid values:

opened

pending-customer-action

reopened

resolved

unassigned

work-in-progress

$sel:caseId:CaseDetails', caseDetails_caseId - The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

$sel:recentCommunications:CaseDetails', caseDetails_recentCommunications - The five most recent communications between you and AWS Support Center, including the IDs of any attachments to the communications. Also includes a nextToken that you can use to retrieve earlier communications.

$sel:serviceCode:CaseDetails', caseDetails_serviceCode - The code for the AWS service. You can get a list of codes and the corresponding service names by calling DescribeServices.

$sel:categoryCode:CaseDetails', caseDetails_categoryCode - The category of problem for the support case.

$sel:submittedBy:CaseDetails', caseDetails_submittedBy - The email address of the account that submitted the case.

$sel:subject:CaseDetails', caseDetails_subject - The subject line for the case in the AWS Support Center.

$sel:ccEmailAddresses:CaseDetails', caseDetails_ccEmailAddresses - The email addresses that receive copies of communication about the case.

$sel:severityCode:CaseDetails', caseDetails_severityCode - The code for the severity level returned by the call to DescribeSeverityLevels.

$sel:timeCreated:CaseDetails', caseDetails_timeCreated - The time that the case was created in the AWS Support Center.

$sel:language:CaseDetails', caseDetails_language - The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

caseDetails_displayId :: Lens' CaseDetails (Maybe Text)

The ID displayed for the case in the AWS Support Center. This is a numeric string.

caseDetails_status :: Lens' CaseDetails (Maybe Text)

The status of the case.

Valid values:

opened

pending-customer-action

reopened

resolved

unassigned

work-in-progress

caseDetails_caseId :: Lens' CaseDetails (Maybe Text)

The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

caseDetails_recentCommunications :: Lens' CaseDetails (Maybe RecentCaseCommunications)

The five most recent communications between you and AWS Support Center, including the IDs of any attachments to the communications. Also includes a nextToken that you can use to retrieve earlier communications.

caseDetails_serviceCode :: Lens' CaseDetails (Maybe Text)

The code for the AWS service. You can get a list of codes and the corresponding service names by calling DescribeServices.

caseDetails_categoryCode :: Lens' CaseDetails (Maybe Text)

The category of problem for the support case.

caseDetails_submittedBy :: Lens' CaseDetails (Maybe Text)

The email address of the account that submitted the case.

caseDetails_subject :: Lens' CaseDetails (Maybe Text)

The subject line for the case in the AWS Support Center.

caseDetails_ccEmailAddresses :: Lens' CaseDetails (Maybe [Text])

The email addresses that receive copies of communication about the case.

caseDetails_severityCode :: Lens' CaseDetails (Maybe Text)

The code for the severity level returned by the call to DescribeSeverityLevels.

caseDetails_timeCreated :: Lens' CaseDetails (Maybe Text)

The time that the case was created in the AWS Support Center.

caseDetails_language :: Lens' CaseDetails (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

Category

data Category

A JSON-formatted name/value pair that represents the category name and category code of the problem, selected from the DescribeServices response for each AWS service.

See: newCategory smart constructor.

Constructors

Category'

Fields

code :: Maybe Text

The category code for the support case.

name :: Maybe Text

The category name for the support case.

newCategory :: Category

Create a value of Category with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:code:Category', category_code - The category code for the support case.

$sel:name:Category', category_name - The category name for the support case.

category_code :: Lens' Category (Maybe Text)

The category code for the support case.

category_name :: Lens' Category (Maybe Text)

The category name for the support case.

Communication

data Communication

A communication associated with a support case. The communication consists of the case ID, the message body, attachment information, the submitter of the communication, and the date and time of the communication.

See: newCommunication smart constructor.

Constructors

Communication'

Fields

caseId :: Maybe Text

The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

attachmentSet :: Maybe [AttachmentDetails]

Information about the attachments to the case communication.

body :: Maybe Text

The text of the communication between the customer and AWS Support.

submittedBy :: Maybe Text

The identity of the account that submitted, or responded to, the support case. Customer entries include the role or IAM user as well as the email address. For example, "AdminRole (Role) <janedoe@example.com>. Entries from the AWS Support team display "Amazon Web Services," and don't show an email address.

timeCreated :: Maybe Text

The time the communication was created.

newCommunication :: Communication

Create a value of Communication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:caseId:Communication', communication_caseId - The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

$sel:attachmentSet:Communication', communication_attachmentSet - Information about the attachments to the case communication.

$sel:body:Communication', communication_body - The text of the communication between the customer and AWS Support.

$sel:submittedBy:Communication', communication_submittedBy - The identity of the account that submitted, or responded to, the support case. Customer entries include the role or IAM user as well as the email address. For example, "AdminRole (Role) <janedoe@example.com>. Entries from the AWS Support team display "Amazon Web Services," and don't show an email address.

$sel:timeCreated:Communication', communication_timeCreated - The time the communication was created.

communication_caseId :: Lens' Communication (Maybe Text)

The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

communication_attachmentSet :: Lens' Communication (Maybe [AttachmentDetails])

Information about the attachments to the case communication.

communication_body :: Lens' Communication (Maybe Text)

The text of the communication between the customer and AWS Support.

communication_submittedBy :: Lens' Communication (Maybe Text)

The identity of the account that submitted, or responded to, the support case. Customer entries include the role or IAM user as well as the email address. For example, "AdminRole (Role) <janedoe@example.com>. Entries from the AWS Support team display "Amazon Web Services," and don't show an email address.

communication_timeCreated :: Lens' Communication (Maybe Text)

The time the communication was created.

RecentCaseCommunications

data RecentCaseCommunications

The five most recent communications associated with the case.

See: newRecentCaseCommunications smart constructor.

Constructors

RecentCaseCommunications'

Fields

nextToken :: Maybe Text

A resumption point for pagination.

communications :: Maybe [Communication]

The five most recent communications associated with the case.

newRecentCaseCommunications :: RecentCaseCommunications

Create a value of RecentCaseCommunications with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:RecentCaseCommunications', recentCaseCommunications_nextToken - A resumption point for pagination.

$sel:communications:RecentCaseCommunications', recentCaseCommunications_communications - The five most recent communications associated with the case.

recentCaseCommunications_nextToken :: Lens' RecentCaseCommunications (Maybe Text)

A resumption point for pagination.

recentCaseCommunications_communications :: Lens' RecentCaseCommunications (Maybe [Communication])

The five most recent communications associated with the case.

SeverityLevel

data SeverityLevel

A code and name pair that represents the severity level of a support case. The available values depend on the support plan for the account. For more information, see Choosing a severity in the AWS Support User Guide.

See: newSeverityLevel smart constructor.

Constructors

SeverityLevel'

Fields

code :: Maybe Text

The code for case severity level.

Valid values: low | normal | high | urgent | critical

name :: Maybe Text

The name of the severity level that corresponds to the severity level code.

The values returned by the API are different from the values that appear in the AWS Support Center. For example, the API uses the code low, but the name appears as General guidance in Support Center.

The following are the API code names and how they appear in the console:

low - General guidance

normal - System impaired

high - Production system impaired

urgent - Production system down

critical - Business-critical system down

For more information, see Choosing a severity in the AWS Support User Guide.

newSeverityLevel :: SeverityLevel

Create a value of SeverityLevel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:code:SeverityLevel', severityLevel_code - The code for case severity level.

Valid values: low | normal | high | urgent | critical

$sel:name:SeverityLevel', severityLevel_name - The name of the severity level that corresponds to the severity level code.

The values returned by the API are different from the values that appear in the AWS Support Center. For example, the API uses the code low, but the name appears as General guidance in Support Center.

The following are the API code names and how they appear in the console:

low - General guidance

normal - System impaired

high - Production system impaired

urgent - Production system down

critical - Business-critical system down

For more information, see Choosing a severity in the AWS Support User Guide.

severityLevel_code :: Lens' SeverityLevel (Maybe Text)

The code for case severity level.

Valid values: low | normal | high | urgent | critical

severityLevel_name :: Lens' SeverityLevel (Maybe Text)

The name of the severity level that corresponds to the severity level code.

The values returned by the API are different from the values that appear in the AWS Support Center. For example, the API uses the code low, but the name appears as General guidance in Support Center.

The following are the API code names and how they appear in the console:

low - General guidance

normal - System impaired

high - Production system impaired

urgent - Production system down

critical - Business-critical system down

For more information, see Choosing a severity in the AWS Support User Guide.

SupportService

data SupportService

Information about an AWS service returned by the DescribeServices operation.

See: newSupportService smart constructor.

Constructors

SupportService'

Fields

code :: Maybe Text

The code for an AWS service returned by the DescribeServices response. The name element contains the corresponding friendly name.

name :: Maybe Text

The friendly name for an AWS service. The code element contains the corresponding code.

categories :: Maybe [Category]

A list of categories that describe the type of support issue a case describes. Categories consist of a category name and a category code. Category names and codes are passed to AWS Support when you call CreateCase.

newSupportService :: SupportService

Create a value of SupportService with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:code:SupportService', supportService_code - The code for an AWS service returned by the DescribeServices response. The name element contains the corresponding friendly name.

$sel:name:SupportService', supportService_name - The friendly name for an AWS service. The code element contains the corresponding code.

$sel:categories:SupportService', supportService_categories - A list of categories that describe the type of support issue a case describes. Categories consist of a category name and a category code. Category names and codes are passed to AWS Support when you call CreateCase.

supportService_code :: Lens' SupportService (Maybe Text)

The code for an AWS service returned by the DescribeServices response. The name element contains the corresponding friendly name.

supportService_name :: Lens' SupportService (Maybe Text)

The friendly name for an AWS service. The code element contains the corresponding code.

supportService_categories :: Lens' SupportService (Maybe [Category])

A list of categories that describe the type of support issue a case describes. Categories consist of a category name and a category code. Category names and codes are passed to AWS Support when you call CreateCase.

TrustedAdvisorCategorySpecificSummary

data TrustedAdvisorCategorySpecificSummary

The container for summary information that relates to the category of the Trusted Advisor check.

See: newTrustedAdvisorCategorySpecificSummary smart constructor.

Constructors

TrustedAdvisorCategorySpecificSummary'

Fields

costOptimizing :: Maybe TrustedAdvisorCostOptimizingSummary

The summary information about cost savings for a Trusted Advisor check that is in the Cost Optimizing category.

newTrustedAdvisorCategorySpecificSummary :: TrustedAdvisorCategorySpecificSummary

Create a value of TrustedAdvisorCategorySpecificSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:costOptimizing:TrustedAdvisorCategorySpecificSummary', trustedAdvisorCategorySpecificSummary_costOptimizing - The summary information about cost savings for a Trusted Advisor check that is in the Cost Optimizing category.

trustedAdvisorCategorySpecificSummary_costOptimizing :: Lens' TrustedAdvisorCategorySpecificSummary (Maybe TrustedAdvisorCostOptimizingSummary)

The summary information about cost savings for a Trusted Advisor check that is in the Cost Optimizing category.

TrustedAdvisorCheckDescription

data TrustedAdvisorCheckDescription

The description and metadata for a Trusted Advisor check.

See: newTrustedAdvisorCheckDescription smart constructor.

Constructors

TrustedAdvisorCheckDescription'

Fields

id :: Text

The unique identifier for the Trusted Advisor check.

name :: Text

The display name for the Trusted Advisor check.

description :: Text

The description of the Trusted Advisor check, which includes the alert criteria and recommended operations (contains HTML markup).

category :: Text

The category of the Trusted Advisor check.

metadata :: [Text]

The column headings for the data returned by the Trusted Advisor check. The order of the headings corresponds to the order of the data in the Metadata element of the TrustedAdvisorResourceDetail for the check. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

newTrustedAdvisorCheckDescription

Arguments

:: Text
$sel:id:TrustedAdvisorCheckDescription'

-> Text
$sel:name:TrustedAdvisorCheckDescription'

-> Text
$sel:description:TrustedAdvisorCheckDescription'

-> Text
$sel:category:TrustedAdvisorCheckDescription'

-> TrustedAdvisorCheckDescription

Create a value of TrustedAdvisorCheckDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:TrustedAdvisorCheckDescription', trustedAdvisorCheckDescription_id - The unique identifier for the Trusted Advisor check.

$sel:name:TrustedAdvisorCheckDescription', trustedAdvisorCheckDescription_name - The display name for the Trusted Advisor check.

$sel:description:TrustedAdvisorCheckDescription', trustedAdvisorCheckDescription_description - The description of the Trusted Advisor check, which includes the alert criteria and recommended operations (contains HTML markup).

$sel:category:TrustedAdvisorCheckDescription', trustedAdvisorCheckDescription_category - The category of the Trusted Advisor check.

$sel:metadata:TrustedAdvisorCheckDescription', trustedAdvisorCheckDescription_metadata - The column headings for the data returned by the Trusted Advisor check. The order of the headings corresponds to the order of the data in the Metadata element of the TrustedAdvisorResourceDetail for the check. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

trustedAdvisorCheckDescription_id :: Lens' TrustedAdvisorCheckDescription Text

The unique identifier for the Trusted Advisor check.

trustedAdvisorCheckDescription_name :: Lens' TrustedAdvisorCheckDescription Text

The display name for the Trusted Advisor check.

trustedAdvisorCheckDescription_description :: Lens' TrustedAdvisorCheckDescription Text

The description of the Trusted Advisor check, which includes the alert criteria and recommended operations (contains HTML markup).

trustedAdvisorCheckDescription_category :: Lens' TrustedAdvisorCheckDescription Text

The category of the Trusted Advisor check.

trustedAdvisorCheckDescription_metadata :: Lens' TrustedAdvisorCheckDescription [Text]

The column headings for the data returned by the Trusted Advisor check. The order of the headings corresponds to the order of the data in the Metadata element of the TrustedAdvisorResourceDetail for the check. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

TrustedAdvisorCheckRefreshStatus

data TrustedAdvisorCheckRefreshStatus

The refresh status of a Trusted Advisor check.

See: newTrustedAdvisorCheckRefreshStatus smart constructor.

Constructors

TrustedAdvisorCheckRefreshStatus'

Fields

checkId :: Text

The unique identifier for the Trusted Advisor check.

status :: Text

The status of the Trusted Advisor check for which a refresh has been requested:

none - The check is not refreshed or the non-success status exceeds the timeout

enqueued - The check refresh requests has entered the refresh queue

processing - The check refresh request is picked up by the rule processing engine

success - The check is successfully refreshed

abandoned - The check refresh has failed

millisUntilNextRefreshable :: Integer

The amount of time, in milliseconds, until the Trusted Advisor check is eligible for refresh.

newTrustedAdvisorCheckRefreshStatus

Arguments

:: Text
$sel:checkId:TrustedAdvisorCheckRefreshStatus'

-> Text
$sel:status:TrustedAdvisorCheckRefreshStatus'

-> Integer
$sel:millisUntilNextRefreshable:TrustedAdvisorCheckRefreshStatus'

-> TrustedAdvisorCheckRefreshStatus

Create a value of TrustedAdvisorCheckRefreshStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checkId:TrustedAdvisorCheckRefreshStatus', trustedAdvisorCheckRefreshStatus_checkId - The unique identifier for the Trusted Advisor check.

$sel:status:TrustedAdvisorCheckRefreshStatus', trustedAdvisorCheckRefreshStatus_status - The status of the Trusted Advisor check for which a refresh has been requested:

none - The check is not refreshed or the non-success status exceeds the timeout

enqueued - The check refresh requests has entered the refresh queue

processing - The check refresh request is picked up by the rule processing engine

success - The check is successfully refreshed

abandoned - The check refresh has failed

$sel:millisUntilNextRefreshable:TrustedAdvisorCheckRefreshStatus', trustedAdvisorCheckRefreshStatus_millisUntilNextRefreshable - The amount of time, in milliseconds, until the Trusted Advisor check is eligible for refresh.

trustedAdvisorCheckRefreshStatus_checkId :: Lens' TrustedAdvisorCheckRefreshStatus Text

The unique identifier for the Trusted Advisor check.

trustedAdvisorCheckRefreshStatus_status :: Lens' TrustedAdvisorCheckRefreshStatus Text

The status of the Trusted Advisor check for which a refresh has been requested:

none - The check is not refreshed or the non-success status exceeds the timeout

enqueued - The check refresh requests has entered the refresh queue

processing - The check refresh request is picked up by the rule processing engine

success - The check is successfully refreshed

abandoned - The check refresh has failed

trustedAdvisorCheckRefreshStatus_millisUntilNextRefreshable :: Lens' TrustedAdvisorCheckRefreshStatus Integer

The amount of time, in milliseconds, until the Trusted Advisor check is eligible for refresh.

TrustedAdvisorCheckResult

data TrustedAdvisorCheckResult

The results of a Trusted Advisor check returned by DescribeTrustedAdvisorCheckResult.

See: newTrustedAdvisorCheckResult smart constructor.

Constructors

TrustedAdvisorCheckResult'

Fields

checkId :: Text

The unique identifier for the Trusted Advisor check.

timestamp :: Text

The time of the last refresh of the check.

status :: Text

The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

resourcesSummary :: TrustedAdvisorResourcesSummary

categorySpecificSummary :: TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

flaggedResources :: [TrustedAdvisorResourceDetail]

The details about each resource listed in the check result.

newTrustedAdvisorCheckResult

Arguments

:: Text
$sel:checkId:TrustedAdvisorCheckResult'

-> Text
$sel:timestamp:TrustedAdvisorCheckResult'

-> Text
$sel:status:TrustedAdvisorCheckResult'

-> TrustedAdvisorResourcesSummary
$sel:resourcesSummary:TrustedAdvisorCheckResult'

-> TrustedAdvisorCategorySpecificSummary
$sel:categorySpecificSummary:TrustedAdvisorCheckResult'

-> TrustedAdvisorCheckResult

Create a value of TrustedAdvisorCheckResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checkId:TrustedAdvisorCheckResult', trustedAdvisorCheckResult_checkId - The unique identifier for the Trusted Advisor check.

$sel:timestamp:TrustedAdvisorCheckResult', trustedAdvisorCheckResult_timestamp - The time of the last refresh of the check.

$sel:status:TrustedAdvisorCheckResult', trustedAdvisorCheckResult_status - The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

$sel:resourcesSummary:TrustedAdvisorCheckResult', trustedAdvisorCheckResult_resourcesSummary - Undocumented member.

$sel:categorySpecificSummary:TrustedAdvisorCheckResult', trustedAdvisorCheckResult_categorySpecificSummary - Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

$sel:flaggedResources:TrustedAdvisorCheckResult', trustedAdvisorCheckResult_flaggedResources - The details about each resource listed in the check result.

trustedAdvisorCheckResult_checkId :: Lens' TrustedAdvisorCheckResult Text

The unique identifier for the Trusted Advisor check.

trustedAdvisorCheckResult_timestamp :: Lens' TrustedAdvisorCheckResult Text

The time of the last refresh of the check.

trustedAdvisorCheckResult_status :: Lens' TrustedAdvisorCheckResult Text

The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

trustedAdvisorCheckResult_resourcesSummary :: Lens' TrustedAdvisorCheckResult TrustedAdvisorResourcesSummary

Undocumented member.

trustedAdvisorCheckResult_categorySpecificSummary :: Lens' TrustedAdvisorCheckResult TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

trustedAdvisorCheckResult_flaggedResources :: Lens' TrustedAdvisorCheckResult [TrustedAdvisorResourceDetail]

The details about each resource listed in the check result.

TrustedAdvisorCheckSummary

data TrustedAdvisorCheckSummary

A summary of a Trusted Advisor check result, including the alert status, last refresh, and number of resources examined.

See: newTrustedAdvisorCheckSummary smart constructor.

Constructors

TrustedAdvisorCheckSummary'

Fields

hasFlaggedResources :: Maybe Bool

Specifies whether the Trusted Advisor check has flagged resources.

checkId :: Text

The unique identifier for the Trusted Advisor check.

timestamp :: Text

The time of the last refresh of the check.

status :: Text

The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

resourcesSummary :: TrustedAdvisorResourcesSummary

categorySpecificSummary :: TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

newTrustedAdvisorCheckSummary

Arguments

:: Text
$sel:checkId:TrustedAdvisorCheckSummary'

-> Text
$sel:timestamp:TrustedAdvisorCheckSummary'

-> Text
$sel:status:TrustedAdvisorCheckSummary'

-> TrustedAdvisorResourcesSummary
$sel:resourcesSummary:TrustedAdvisorCheckSummary'

-> TrustedAdvisorCategorySpecificSummary
$sel:categorySpecificSummary:TrustedAdvisorCheckSummary'

-> TrustedAdvisorCheckSummary

Create a value of TrustedAdvisorCheckSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hasFlaggedResources:TrustedAdvisorCheckSummary', trustedAdvisorCheckSummary_hasFlaggedResources - Specifies whether the Trusted Advisor check has flagged resources.

$sel:checkId:TrustedAdvisorCheckSummary', trustedAdvisorCheckSummary_checkId - The unique identifier for the Trusted Advisor check.

$sel:timestamp:TrustedAdvisorCheckSummary', trustedAdvisorCheckSummary_timestamp - The time of the last refresh of the check.

$sel:status:TrustedAdvisorCheckSummary', trustedAdvisorCheckSummary_status - The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

$sel:resourcesSummary:TrustedAdvisorCheckSummary', trustedAdvisorCheckSummary_resourcesSummary - Undocumented member.

$sel:categorySpecificSummary:TrustedAdvisorCheckSummary', trustedAdvisorCheckSummary_categorySpecificSummary - Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

trustedAdvisorCheckSummary_hasFlaggedResources :: Lens' TrustedAdvisorCheckSummary (Maybe Bool)

Specifies whether the Trusted Advisor check has flagged resources.

trustedAdvisorCheckSummary_checkId :: Lens' TrustedAdvisorCheckSummary Text

The unique identifier for the Trusted Advisor check.

trustedAdvisorCheckSummary_timestamp :: Lens' TrustedAdvisorCheckSummary Text

The time of the last refresh of the check.

trustedAdvisorCheckSummary_status :: Lens' TrustedAdvisorCheckSummary Text

The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

trustedAdvisorCheckSummary_resourcesSummary :: Lens' TrustedAdvisorCheckSummary TrustedAdvisorResourcesSummary

Undocumented member.

trustedAdvisorCheckSummary_categorySpecificSummary :: Lens' TrustedAdvisorCheckSummary TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

TrustedAdvisorCostOptimizingSummary

data TrustedAdvisorCostOptimizingSummary

The estimated cost savings that might be realized if the recommended operations are taken.

See: newTrustedAdvisorCostOptimizingSummary smart constructor.

Constructors

TrustedAdvisorCostOptimizingSummary'

Fields

estimatedMonthlySavings :: Double

The estimated monthly savings that might be realized if the recommended operations are taken.

estimatedPercentMonthlySavings :: Double

The estimated percentage of savings that might be realized if the recommended operations are taken.

newTrustedAdvisorCostOptimizingSummary

Arguments

:: Double
$sel:estimatedMonthlySavings:TrustedAdvisorCostOptimizingSummary'

-> Double
$sel:estimatedPercentMonthlySavings:TrustedAdvisorCostOptimizingSummary'

-> TrustedAdvisorCostOptimizingSummary

Create a value of TrustedAdvisorCostOptimizingSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:estimatedMonthlySavings:TrustedAdvisorCostOptimizingSummary', trustedAdvisorCostOptimizingSummary_estimatedMonthlySavings - The estimated monthly savings that might be realized if the recommended operations are taken.

$sel:estimatedPercentMonthlySavings:TrustedAdvisorCostOptimizingSummary', trustedAdvisorCostOptimizingSummary_estimatedPercentMonthlySavings - The estimated percentage of savings that might be realized if the recommended operations are taken.

trustedAdvisorCostOptimizingSummary_estimatedMonthlySavings :: Lens' TrustedAdvisorCostOptimizingSummary Double

The estimated monthly savings that might be realized if the recommended operations are taken.

trustedAdvisorCostOptimizingSummary_estimatedPercentMonthlySavings :: Lens' TrustedAdvisorCostOptimizingSummary Double

The estimated percentage of savings that might be realized if the recommended operations are taken.

TrustedAdvisorResourceDetail

data TrustedAdvisorResourceDetail

Contains information about a resource identified by a Trusted Advisor check.

See: newTrustedAdvisorResourceDetail smart constructor.

Constructors

TrustedAdvisorResourceDetail'

Fields

isSuppressed :: Maybe Bool

Specifies whether the AWS resource was ignored by Trusted Advisor because it was marked as suppressed by the user.

region :: Maybe Text

The AWS Region in which the identified resource is located.

status :: Text

The status code for the resource identified in the Trusted Advisor check.

resourceId :: Text

The unique identifier for the identified resource.

metadata :: [Text]

Additional information about the identified resource. The exact metadata and its order can be obtained by inspecting the TrustedAdvisorCheckDescription object returned by the call to DescribeTrustedAdvisorChecks. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

newTrustedAdvisorResourceDetail

Arguments

:: Text
$sel:status:TrustedAdvisorResourceDetail'

-> Text
$sel:resourceId:TrustedAdvisorResourceDetail'

-> TrustedAdvisorResourceDetail

Create a value of TrustedAdvisorResourceDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isSuppressed:TrustedAdvisorResourceDetail', trustedAdvisorResourceDetail_isSuppressed - Specifies whether the AWS resource was ignored by Trusted Advisor because it was marked as suppressed by the user.

$sel:region:TrustedAdvisorResourceDetail', trustedAdvisorResourceDetail_region - The AWS Region in which the identified resource is located.

$sel:status:TrustedAdvisorResourceDetail', trustedAdvisorResourceDetail_status - The status code for the resource identified in the Trusted Advisor check.

$sel:resourceId:TrustedAdvisorResourceDetail', trustedAdvisorResourceDetail_resourceId - The unique identifier for the identified resource.

$sel:metadata:TrustedAdvisorResourceDetail', trustedAdvisorResourceDetail_metadata - Additional information about the identified resource. The exact metadata and its order can be obtained by inspecting the TrustedAdvisorCheckDescription object returned by the call to DescribeTrustedAdvisorChecks. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

trustedAdvisorResourceDetail_isSuppressed :: Lens' TrustedAdvisorResourceDetail (Maybe Bool)

Specifies whether the AWS resource was ignored by Trusted Advisor because it was marked as suppressed by the user.

trustedAdvisorResourceDetail_region :: Lens' TrustedAdvisorResourceDetail (Maybe Text)

The AWS Region in which the identified resource is located.

trustedAdvisorResourceDetail_status :: Lens' TrustedAdvisorResourceDetail Text

The status code for the resource identified in the Trusted Advisor check.

trustedAdvisorResourceDetail_resourceId :: Lens' TrustedAdvisorResourceDetail Text

The unique identifier for the identified resource.

trustedAdvisorResourceDetail_metadata :: Lens' TrustedAdvisorResourceDetail [Text]

Additional information about the identified resource. The exact metadata and its order can be obtained by inspecting the TrustedAdvisorCheckDescription object returned by the call to DescribeTrustedAdvisorChecks. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

TrustedAdvisorResourcesSummary

data TrustedAdvisorResourcesSummary

Details about AWS resources that were analyzed in a call to Trusted Advisor DescribeTrustedAdvisorCheckSummaries.

See: newTrustedAdvisorResourcesSummary smart constructor.

Constructors

TrustedAdvisorResourcesSummary'

Fields

resourcesProcessed :: Integer

The number of AWS resources that were analyzed by the Trusted Advisor check.

resourcesFlagged :: Integer

The number of AWS resources that were flagged (listed) by the Trusted Advisor check.

resourcesIgnored :: Integer

The number of AWS resources ignored by Trusted Advisor because information was unavailable.

resourcesSuppressed :: Integer

The number of AWS resources ignored by Trusted Advisor because they were marked as suppressed by the user.

newTrustedAdvisorResourcesSummary

Arguments

:: Integer
$sel:resourcesProcessed:TrustedAdvisorResourcesSummary'

-> Integer
$sel:resourcesFlagged:TrustedAdvisorResourcesSummary'

-> Integer
$sel:resourcesIgnored:TrustedAdvisorResourcesSummary'

-> Integer
$sel:resourcesSuppressed:TrustedAdvisorResourcesSummary'

-> TrustedAdvisorResourcesSummary

Create a value of TrustedAdvisorResourcesSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourcesProcessed:TrustedAdvisorResourcesSummary', trustedAdvisorResourcesSummary_resourcesProcessed - The number of AWS resources that were analyzed by the Trusted Advisor check.

$sel:resourcesFlagged:TrustedAdvisorResourcesSummary', trustedAdvisorResourcesSummary_resourcesFlagged - The number of AWS resources that were flagged (listed) by the Trusted Advisor check.

$sel:resourcesIgnored:TrustedAdvisorResourcesSummary', trustedAdvisorResourcesSummary_resourcesIgnored - The number of AWS resources ignored by Trusted Advisor because information was unavailable.

$sel:resourcesSuppressed:TrustedAdvisorResourcesSummary', trustedAdvisorResourcesSummary_resourcesSuppressed - The number of AWS resources ignored by Trusted Advisor because they were marked as suppressed by the user.

trustedAdvisorResourcesSummary_resourcesProcessed :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources that were analyzed by the Trusted Advisor check.

trustedAdvisorResourcesSummary_resourcesFlagged :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources that were flagged (listed) by the Trusted Advisor check.

trustedAdvisorResourcesSummary_resourcesIgnored :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources ignored by Trusted Advisor because information was unavailable.

trustedAdvisorResourcesSummary_resourcesSuppressed :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources ignored by Trusted Advisor because they were marked as suppressed by the user.