Network.AWS.Support.Types.SeverityLevel

Description

Documentation

data SeverityLevel

A code and name pair that represents the severity level of a support case. The available values depend on the support plan for the account. For more information, see Choosing a severity in the AWS Support User Guide.

See: newSeverityLevel smart constructor.

Constructors

SeverityLevel'

Fields

name :: Maybe Text

The name of the severity level that corresponds to the severity level code.

The values returned by the API are different from the values that appear in the AWS Support Center. For example, the API uses the code low, but the name appears as General guidance in Support Center.

The following are the API code names and how they appear in the console:

low - General guidance

normal - System impaired

high - Production system impaired

urgent - Production system down

critical - Business-critical system down

For more information, see Choosing a severity in the AWS Support User Guide.

code :: Maybe Text

The code for case severity level.

Valid values: low | normal | high | urgent | critical

newSeverityLevel :: SeverityLevel

Create a value of SeverityLevel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:SeverityLevel', severityLevel_name - The name of the severity level that corresponds to the severity level code.

The values returned by the API are different from the values that appear in the AWS Support Center. For example, the API uses the code low, but the name appears as General guidance in Support Center.

The following are the API code names and how they appear in the console:

low - General guidance

normal - System impaired

high - Production system impaired

urgent - Production system down

critical - Business-critical system down

For more information, see Choosing a severity in the AWS Support User Guide.

$sel:code:SeverityLevel', severityLevel_code - The code for case severity level.

Valid values: low | normal | high | urgent | critical

severityLevel_name :: Lens' SeverityLevel (Maybe Text)

The name of the severity level that corresponds to the severity level code.

The values returned by the API are different from the values that appear in the AWS Support Center. For example, the API uses the code low, but the name appears as General guidance in Support Center.

The following are the API code names and how they appear in the console:

low - General guidance

normal - System impaired

high - Production system impaired

urgent - Production system down

critical - Business-critical system down

For more information, see Choosing a severity in the AWS Support User Guide.

severityLevel_code :: Lens' SeverityLevel (Maybe Text)

The code for case severity level.

Valid values: low | normal | high | urgent | critical