Network.AWS.Support.Types.TrustedAdvisorCheckDescription

Description

Documentation

data TrustedAdvisorCheckDescription

The description and metadata for a Trusted Advisor check.

See: newTrustedAdvisorCheckDescription smart constructor.

Constructors

TrustedAdvisorCheckDescription'

Fields

id :: Text

The unique identifier for the Trusted Advisor check.

name :: Text

The display name for the Trusted Advisor check.

description :: Text

The description of the Trusted Advisor check, which includes the alert criteria and recommended operations (contains HTML markup).

category :: Text

The category of the Trusted Advisor check.

metadata :: [Text]

The column headings for the data returned by the Trusted Advisor check. The order of the headings corresponds to the order of the data in the Metadata element of the TrustedAdvisorResourceDetail for the check. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

newTrustedAdvisorCheckDescription

Arguments

:: Text
$sel:id:TrustedAdvisorCheckDescription'

-> Text
$sel:name:TrustedAdvisorCheckDescription'

-> Text
$sel:description:TrustedAdvisorCheckDescription'

-> Text
$sel:category:TrustedAdvisorCheckDescription'

-> TrustedAdvisorCheckDescription

Create a value of TrustedAdvisorCheckDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:TrustedAdvisorCheckDescription', trustedAdvisorCheckDescription_id - The unique identifier for the Trusted Advisor check.

$sel:name:TrustedAdvisorCheckDescription', trustedAdvisorCheckDescription_name - The display name for the Trusted Advisor check.

$sel:description:TrustedAdvisorCheckDescription', trustedAdvisorCheckDescription_description - The description of the Trusted Advisor check, which includes the alert criteria and recommended operations (contains HTML markup).

$sel:category:TrustedAdvisorCheckDescription', trustedAdvisorCheckDescription_category - The category of the Trusted Advisor check.

$sel:metadata:TrustedAdvisorCheckDescription', trustedAdvisorCheckDescription_metadata - The column headings for the data returned by the Trusted Advisor check. The order of the headings corresponds to the order of the data in the Metadata element of the TrustedAdvisorResourceDetail for the check. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

trustedAdvisorCheckDescription_id :: Lens' TrustedAdvisorCheckDescription Text

The unique identifier for the Trusted Advisor check.

trustedAdvisorCheckDescription_name :: Lens' TrustedAdvisorCheckDescription Text

The display name for the Trusted Advisor check.

trustedAdvisorCheckDescription_description :: Lens' TrustedAdvisorCheckDescription Text

The description of the Trusted Advisor check, which includes the alert criteria and recommended operations (contains HTML markup).

trustedAdvisorCheckDescription_category :: Lens' TrustedAdvisorCheckDescription Text

The category of the Trusted Advisor check.

trustedAdvisorCheckDescription_metadata :: Lens' TrustedAdvisorCheckDescription [Text]

The column headings for the data returned by the Trusted Advisor check. The order of the headings corresponds to the order of the data in the Metadata element of the TrustedAdvisorResourceDetail for the check. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.