Network.AWS.Support.Types.TrustedAdvisorCheckRefreshStatus

Description

Documentation

data TrustedAdvisorCheckRefreshStatus

The refresh status of a Trusted Advisor check.

See: newTrustedAdvisorCheckRefreshStatus smart constructor.

Constructors

TrustedAdvisorCheckRefreshStatus'

Fields

checkId :: Text

The unique identifier for the Trusted Advisor check.

status :: Text

The status of the Trusted Advisor check for which a refresh has been requested:

none - The check is not refreshed or the non-success status exceeds the timeout

enqueued - The check refresh requests has entered the refresh queue

processing - The check refresh request is picked up by the rule processing engine

success - The check is successfully refreshed

abandoned - The check refresh has failed

millisUntilNextRefreshable :: Integer

The amount of time, in milliseconds, until the Trusted Advisor check is eligible for refresh.

newTrustedAdvisorCheckRefreshStatus

Arguments

:: Text
$sel:checkId:TrustedAdvisorCheckRefreshStatus'

-> Text
$sel:status:TrustedAdvisorCheckRefreshStatus'

-> Integer
$sel:millisUntilNextRefreshable:TrustedAdvisorCheckRefreshStatus'

-> TrustedAdvisorCheckRefreshStatus

Create a value of TrustedAdvisorCheckRefreshStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checkId:TrustedAdvisorCheckRefreshStatus', trustedAdvisorCheckRefreshStatus_checkId - The unique identifier for the Trusted Advisor check.

$sel:status:TrustedAdvisorCheckRefreshStatus', trustedAdvisorCheckRefreshStatus_status - The status of the Trusted Advisor check for which a refresh has been requested:

none - The check is not refreshed or the non-success status exceeds the timeout

enqueued - The check refresh requests has entered the refresh queue

processing - The check refresh request is picked up by the rule processing engine

success - The check is successfully refreshed

abandoned - The check refresh has failed

$sel:millisUntilNextRefreshable:TrustedAdvisorCheckRefreshStatus', trustedAdvisorCheckRefreshStatus_millisUntilNextRefreshable - The amount of time, in milliseconds, until the Trusted Advisor check is eligible for refresh.

trustedAdvisorCheckRefreshStatus_checkId :: Lens' TrustedAdvisorCheckRefreshStatus Text

The unique identifier for the Trusted Advisor check.

trustedAdvisorCheckRefreshStatus_status :: Lens' TrustedAdvisorCheckRefreshStatus Text

The status of the Trusted Advisor check for which a refresh has been requested:

none - The check is not refreshed or the non-success status exceeds the timeout

enqueued - The check refresh requests has entered the refresh queue

processing - The check refresh request is picked up by the rule processing engine

success - The check is successfully refreshed

abandoned - The check refresh has failed

trustedAdvisorCheckRefreshStatus_millisUntilNextRefreshable :: Lens' TrustedAdvisorCheckRefreshStatus Integer

The amount of time, in milliseconds, until the Trusted Advisor check is eligible for refresh.