Network.AWS.Support.Types.TrustedAdvisorResourcesSummary

Description

Documentation

data TrustedAdvisorResourcesSummary

Details about AWS resources that were analyzed in a call to Trusted Advisor DescribeTrustedAdvisorCheckSummaries.

See: newTrustedAdvisorResourcesSummary smart constructor.

Constructors

TrustedAdvisorResourcesSummary'

Fields

resourcesProcessed :: Integer

The number of AWS resources that were analyzed by the Trusted Advisor check.

resourcesFlagged :: Integer

The number of AWS resources that were flagged (listed) by the Trusted Advisor check.

resourcesIgnored :: Integer

The number of AWS resources ignored by Trusted Advisor because information was unavailable.

resourcesSuppressed :: Integer

The number of AWS resources ignored by Trusted Advisor because they were marked as suppressed by the user.

newTrustedAdvisorResourcesSummary

Arguments

:: Integer
$sel:resourcesProcessed:TrustedAdvisorResourcesSummary'

-> Integer
$sel:resourcesFlagged:TrustedAdvisorResourcesSummary'

-> Integer
$sel:resourcesIgnored:TrustedAdvisorResourcesSummary'

-> Integer
$sel:resourcesSuppressed:TrustedAdvisorResourcesSummary'

-> TrustedAdvisorResourcesSummary

Create a value of TrustedAdvisorResourcesSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourcesProcessed:TrustedAdvisorResourcesSummary', trustedAdvisorResourcesSummary_resourcesProcessed - The number of AWS resources that were analyzed by the Trusted Advisor check.

$sel:resourcesFlagged:TrustedAdvisorResourcesSummary', trustedAdvisorResourcesSummary_resourcesFlagged - The number of AWS resources that were flagged (listed) by the Trusted Advisor check.

$sel:resourcesIgnored:TrustedAdvisorResourcesSummary', trustedAdvisorResourcesSummary_resourcesIgnored - The number of AWS resources ignored by Trusted Advisor because information was unavailable.

$sel:resourcesSuppressed:TrustedAdvisorResourcesSummary', trustedAdvisorResourcesSummary_resourcesSuppressed - The number of AWS resources ignored by Trusted Advisor because they were marked as suppressed by the user.

trustedAdvisorResourcesSummary_resourcesProcessed :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources that were analyzed by the Trusted Advisor check.

trustedAdvisorResourcesSummary_resourcesFlagged :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources that were flagged (listed) by the Trusted Advisor check.

trustedAdvisorResourcesSummary_resourcesIgnored :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources ignored by Trusted Advisor because information was unavailable.

trustedAdvisorResourcesSummary_resourcesSuppressed :: Lens' TrustedAdvisorResourcesSummary Integer

The number of AWS resources ignored by Trusted Advisor because they were marked as suppressed by the user.