Network.AWS.Support.V2013_04_15

Description

AWS Support is a one-on-one, fast-response support channel that is staffed with experienced support personnel. The service helps customers get the most from the products and features provided by Amazon Web Services. There are four levels, or tiers, of AWS Support: Basic, Developer, Business, and Enterprise. The Basic tier is free of charge and offers support for account and billing questions and service limit increases. The other tiers offer an unlimited number of technical support cases with pay-by-the-month pricing and no long-term contracts, providing developers and businesses flexibility to choose the level of support that meets their needs.

Documentation

addAttachmentsToSet

Arguments

:: [Attachment]
_aatsrAttachments

-> AddAttachmentsToSet

Minimum specification for a AddAttachmentsToSet request.

data AddAttachmentsToSet

Constructors

AddAttachmentsToSet

Fields

_aatsrAttachments :: [Attachment]

One or more attachments to add to the set. The limit is 3 attachments per set, and the size limit is 5 MB per attachment.

_aatsrAttachmentSetId :: Maybe Text

The ID of the attachment set. If an AttachmentSetId is not specified, a new attachment set is created, and the ID of the set is returned in the response. If an AttachmentSetId is specified, the attachments are added to the specified set, if it exists.

aatsrAttachments :: Lens' AddAttachmentsToSet [Attachment]

aatsrAttachmentSetId :: Lens' AddAttachmentsToSet (Maybe Text)

data AddAttachmentsToSetResponse

Constructors

AddAttachmentsToSetResponse

Fields

_aatssAttachmentSetId :: Maybe Text

The ID of the attachment set. If an AttachmentSetId was not specified, a new attachment set is created, and the ID of the set is returned in the response. If an AttachmentSetId was specified, the attachments are added to the specified set, if it exists.

_aatssExpiryTime :: Maybe Text

The time and date when the attachment set expires.

aatssExpiryTime :: Lens' AddAttachmentsToSetResponse (Maybe Text)

aatssAttachmentSetId :: Lens' AddAttachmentsToSetResponse (Maybe Text)

addCommunicationToCase

Arguments

:: Text
_actcrCommunicationBody

-> AddCommunicationToCase

Minimum specification for a AddCommunicationToCase request.

data AddCommunicationToCase

Constructors

AddCommunicationToCase

Fields

_actcrCommunicationBody :: Text

The body of an email communication to add to the support case.

_actcrAttachmentSetId :: Maybe Text

The ID of a set of one or more attachments for the communication to add to the case. Create the set by calling AddAttachmentsToSet.

_actcrCaseId :: Maybe Text

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47.

_actcrCcEmailAddresses :: [Text]

The email addresses in the CC line of an email to be added to the support case.

actcrCommunicationBody :: Lens' AddCommunicationToCase Text

actcrCcEmailAddresses :: Lens' AddCommunicationToCase [Text]

actcrCaseId :: Lens' AddCommunicationToCase (Maybe Text)

actcrAttachmentSetId :: Lens' AddCommunicationToCase (Maybe Text)

data AddCommunicationToCaseResponse

Constructors

AddCommunicationToCaseResponse

Fields

_actcsResult :: Maybe Bool

True if AddCommunicationToCase succeeds. Otherwise, returns an error.

actcsResult :: Iso' AddCommunicationToCaseResponse (Maybe Bool)

createCase

Arguments

:: Text
_ccrCommunicationBody

-> Text
_ccrSubject

-> CreateCase

Minimum specification for a CreateCase request.

data CreateCase

Constructors

CreateCase

Fields

_ccrCommunicationBody :: Text

The communication body text when you create an AWS Support case by calling CreateCase.

_ccrSubject :: Text

The title of the AWS Support case.

_ccrAttachmentSetId :: Maybe Text

The ID of a set of one or more attachments for the case. Create the set by using AddAttachmentsToSet.

_ccrCategoryCode :: Maybe Text

The category of problem for the AWS Support case.

_ccrCcEmailAddresses :: [Text]

A list of email addresses that AWS Support copies on case correspondence.

_ccrIssueType :: Maybe Text

The type of issue for the case. You can specify either "customer-service" or "technical." If you do not indicate a value, the default is "technical.".

_ccrLanguage :: Maybe Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

_ccrServiceCode :: Maybe Text

The code for the AWS service returned by the call to DescribeServices.

_ccrSeverityCode :: Maybe Text

The code for the severity level returned by the call to DescribeSeverityLevels. The availability of severity levels depends on each customer's support subscription. In other words, your subscription may not necessarily require the urgent level of response time.

ccrSubject :: Lens' CreateCase Text

ccrSeverityCode :: Lens' CreateCase (Maybe Text)

ccrServiceCode :: Lens' CreateCase (Maybe Text)

ccrLanguage :: Lens' CreateCase (Maybe Text)

ccrIssueType :: Lens' CreateCase (Maybe Text)

ccrCommunicationBody :: Lens' CreateCase Text

ccrCcEmailAddresses :: Lens' CreateCase [Text]

ccrCategoryCode :: Lens' CreateCase (Maybe Text)

ccrAttachmentSetId :: Lens' CreateCase (Maybe Text)

data CreateCaseResponse

Constructors

CreateCaseResponse

Fields

_ccsCaseId :: Maybe Text

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47.

ccsCaseId :: Iso' CreateCaseResponse (Maybe Text)

data DescribeAttachment

Constructors

DescribeAttachment

Fields

_darAttachmentId :: Text

The ID of the attachment to return. Attachment IDs are returned by the DescribeCommunications operation.

darAttachmentId :: Iso' DescribeAttachment Text

data DescribeAttachmentResponse

Constructors

DescribeAttachmentResponse

Fields

_dasAttachment :: Maybe Attachment

The attachment content and file name.

dasAttachment :: Iso' DescribeAttachmentResponse (Maybe Attachment)

describeCases :: DescribeCases

Minimum specification for a DescribeCases request.

data DescribeCases

Constructors

DescribeCases

Fields

_dcrAfterTime :: Maybe Text

The start date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

_dcrBeforeTime :: Maybe Text

The end date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

_dcrCaseIdList :: [Text]

A list of ID numbers of the support cases you want returned. The maximum number of cases is 100.

_dcrDisplayId :: Maybe Text

The ID displayed for a case in the AWS Support Center user interface.

_dcrIncludeCommunications :: Maybe Bool

Specifies whether communications should be included in the DescribeCases results. The default is true.

_dcrIncludeResolvedCases :: Maybe Bool

Specifies whether resolved support cases should be included in the DescribeCases results. The default is false.

_dcrLanguage :: Maybe Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

_dcrMaxResults :: Maybe Integer

The maximum number of results to return before paginating.

_dcrNextToken :: Maybe Text

A resumption point for pagination.

dcrNextToken :: Lens' DescribeCases (Maybe Text)

dcrMaxResults :: Lens' DescribeCases (Maybe Integer)

dcrLanguage :: Lens' DescribeCases (Maybe Text)

dcrIncludeResolvedCases :: Lens' DescribeCases (Maybe Bool)

dcrIncludeCommunications :: Lens' DescribeCases (Maybe Bool)

dcrDisplayId :: Lens' DescribeCases (Maybe Text)

dcrCaseIdList :: Lens' DescribeCases [Text]

dcrBeforeTime :: Lens' DescribeCases (Maybe Text)

dcrAfterTime :: Lens' DescribeCases (Maybe Text)

data DescribeCasesResponse

Constructors

DescribeCasesResponse

Fields

_dcsCases :: [CaseDetails]

The details for the cases that match the request.

_dcsNextToken :: Maybe Text

A resumption point for pagination.

dcsNextToken :: Lens' DescribeCasesResponse (Maybe Text)

dcsCases :: Lens' DescribeCasesResponse [CaseDetails]

describeCommunications

Arguments

:: Text
_dctCaseId

-> DescribeCommunications

Minimum specification for a DescribeCommunications request.

data DescribeCommunications

Constructors

DescribeCommunications

Fields

_dctCaseId :: Text

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47.

_dctAfterTime :: Maybe Text

The start date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

_dctBeforeTime :: Maybe Text

The end date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

_dctMaxResults :: Maybe Integer

The maximum number of results to return before paginating.

_dctNextToken :: Maybe Text

A resumption point for pagination.

dctNextToken :: Lens' DescribeCommunications (Maybe Text)

dctMaxResults :: Lens' DescribeCommunications (Maybe Integer)

dctCaseId :: Lens' DescribeCommunications Text

dctBeforeTime :: Lens' DescribeCommunications (Maybe Text)

dctAfterTime :: Lens' DescribeCommunications (Maybe Text)

data DescribeCommunicationsResponse

Constructors

DescribeCommunicationsResponse

Fields

_dcuCommunications :: [Communication]

The communications for the case.

_dcuNextToken :: Maybe Text

A resumption point for pagination.

dcuNextToken :: Lens' DescribeCommunicationsResponse (Maybe Text)

dcuCommunications :: Lens' DescribeCommunicationsResponse [Communication]

describeServices :: DescribeServices

Minimum specification for a DescribeServices request.

data DescribeServices

Constructors

DescribeServices

Fields

_dsrLanguage :: Maybe Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

_dsrServiceCodeList :: [Text]

A JSON-formatted list of service codes available for AWS services.

dsrServiceCodeList :: Lens' DescribeServices [Text]

dsrLanguage :: Lens' DescribeServices (Maybe Text)

data DescribeServicesResponse

Constructors

DescribeServicesResponse

Fields

_dssServices :: [Service]

A JSON-formatted list of AWS services.

dssServices :: Iso' DescribeServicesResponse [Service]

describeSeverityLevels :: DescribeSeverityLevels

Minimum specification for a DescribeSeverityLevels request.

data DescribeSeverityLevels

Constructors

DescribeSeverityLevels

Fields

_dslrLanguage :: Maybe Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

dslrLanguage :: Iso' DescribeSeverityLevels (Maybe Text)

data DescribeSeverityLevelsResponse

Constructors

DescribeSeverityLevelsResponse

Fields

_dslsSeverityLevels :: [SeverityLevel]

The available severity levels for the support case. Available severity levels are defined by your service level agreement with AWS.

dslsSeverityLevels :: Iso' DescribeSeverityLevelsResponse [SeverityLevel]

data DescribeTrustedAdvisorCheckRefreshStatuses

Constructors

DescribeTrustedAdvisorCheckRefreshStatuses

Fields

_dtacrsrCheckIds :: [Text]

The IDs of the Trusted Advisor checks.

dtacrsrCheckIds :: Iso' DescribeTrustedAdvisorCheckRefreshStatuses [Text]

data DescribeTrustedAdvisorCheckRefreshStatusesResponse

Constructors

DescribeTrustedAdvisorCheckRefreshStatusesResponse

Fields

_dtacrssStatuses :: [TrustedAdvisorCheckRefreshStatus]

The refresh status of the specified Trusted Advisor checks.

dtacrssStatuses :: Iso' DescribeTrustedAdvisorCheckRefreshStatusesResponse [TrustedAdvisorCheckRefreshStatus]

describeTrustedAdvisorCheckResult

Arguments

:: Text
_dtacrrCheckId

-> DescribeTrustedAdvisorCheckResult

Minimum specification for a DescribeTrustedAdvisorCheckResult request.

data DescribeTrustedAdvisorCheckResult

Constructors

DescribeTrustedAdvisorCheckResult

Fields

_dtacrrCheckId :: Text

The unique identifier for the Trusted Advisor check.

_dtacrrLanguage :: Maybe Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

dtacrrLanguage :: Lens' DescribeTrustedAdvisorCheckResult (Maybe Text)

dtacrrCheckId :: Lens' DescribeTrustedAdvisorCheckResult Text

data DescribeTrustedAdvisorCheckResultResponse

Constructors

DescribeTrustedAdvisorCheckResultResponse

Fields

_dtacrsResult :: Maybe TrustedAdvisorCheckResult

The detailed results of the Trusted Advisor check.

dtacrsResult :: Iso' DescribeTrustedAdvisorCheckResultResponse (Maybe TrustedAdvisorCheckResult)

data DescribeTrustedAdvisorCheckSummaries

Constructors

DescribeTrustedAdvisorCheckSummaries

Fields

_dtacsrCheckIds :: [Text]

The IDs of the Trusted Advisor checks.

dtacsrCheckIds :: Iso' DescribeTrustedAdvisorCheckSummaries [Text]

data DescribeTrustedAdvisorCheckSummariesResponse

Constructors

DescribeTrustedAdvisorCheckSummariesResponse

Fields

_dtacssSummaries :: [TrustedAdvisorCheckSummary]

The summary information for the requested Trusted Advisor checks.

dtacssSummaries :: Iso' DescribeTrustedAdvisorCheckSummariesResponse [TrustedAdvisorCheckSummary]

data DescribeTrustedAdvisorChecks

Constructors

DescribeTrustedAdvisorChecks

Fields

_dtacrLanguage :: Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

dtacrLanguage :: Iso' DescribeTrustedAdvisorChecks Text

data DescribeTrustedAdvisorChecksResponse

Constructors

DescribeTrustedAdvisorChecksResponse

Fields

_dtacsChecks :: [TrustedAdvisorCheckDescription]

Information about all available Trusted Advisor checks.

dtacsChecks :: Iso' DescribeTrustedAdvisorChecksResponse [TrustedAdvisorCheckDescription]

data RefreshTrustedAdvisorCheck

Constructors

RefreshTrustedAdvisorCheck

Fields

_rtacrCheckId :: Text

The unique identifier for the Trusted Advisor check.

rtacrCheckId :: Iso' RefreshTrustedAdvisorCheck Text

data RefreshTrustedAdvisorCheckResponse

Constructors

RefreshTrustedAdvisorCheckResponse

Fields

_rtacsStatus :: TrustedAdvisorCheckRefreshStatus

The current refresh status for a check, including the amount of time until the check is eligible for refresh.

rtacsStatus :: Iso' RefreshTrustedAdvisorCheckResponse TrustedAdvisorCheckRefreshStatus

resolveCase :: ResolveCase

Minimum specification for a ResolveCase request.

data ResolveCase

Constructors

ResolveCase

Fields

_rcrCaseId :: Maybe Text

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47.

rcrCaseId :: Iso' ResolveCase (Maybe Text)

data ResolveCaseResponse

Constructors

ResolveCaseResponse

Fields

_rcsInitialCaseStatus :: Maybe Text

The status of the case when the ResolveCase request was sent.

_rcsFinalCaseStatus :: Maybe Text

The status of the case after the ResolveCase request was processed.

rcsInitialCaseStatus :: Lens' ResolveCaseResponse (Maybe Text)

rcsFinalCaseStatus :: Lens' ResolveCaseResponse (Maybe Text)

data Support

Supported version (2013-04-15) of the AWS Support service.

newtype TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

Constructors

TrustedAdvisorCategorySpecificSummary

Fields

_tacssCostOptimizing :: Maybe TrustedAdvisorCostOptimizingSummary

The summary information about cost savings for a Trusted Advisor check that is in the Cost Optimizing category.

data Attachment

The attachment content and file name.

Constructors

Attachment

Fields

_aData :: Maybe Base64

The content of the attachment file.

_aFileName :: Maybe Text

The name of the attachment file.

data AttachmentDetails

The file name and ID of an attachment to a case communication. You can use the ID to retrieve the attachment with the DescribeAttachment operation.

Constructors

AttachmentDetails

Fields

_adAttachmentId :: Maybe Text

The ID of the attachment.

_adFileName :: Maybe Text

The file name of the attachment.

data CaseDetails

A JSON-formatted object that contains the metadata for a support case. It is contained the response from a DescribeCases request. CaseDetails contains the following fields: CaseID. The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47. CategoryCode. The category of problem for the AWS Support case. Corresponds to the CategoryCode values returned by a call to DescribeServices. DisplayId. The identifier for the case on pages in the AWS Support Center. Language. The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them. RecentCommunications. One or more Communication objects. Fields of these objects are Attachments, Body, CaseId, SubmittedBy, and TimeCreated. NextToken. A resumption point for pagination. ServiceCode. The identifier for the AWS service that corresponds to the service code defined in the call to DescribeServices. SeverityCode. The severity code assigned to the case. Contains one of the values returned by the call to DescribeSeverityLevels. Status. The status of the case in the AWS Support Center. Subject. The subject line of the case. SubmittedBy. The email address of the account that submitted the case. TimeCreated. The time the case was created, in ISO-8601 format.

Constructors

CaseDetails

Fields

_cdSubject :: Maybe Text

The subject line for the case in the AWS Support Center.

_cdStatus :: Maybe Text

The status of the case.

_cdRecentCommunications :: Maybe RecentCaseCommunications

The five most recent communications between you and AWS Support Center, including the IDs of any attachments to the communications. Also includes a nextToken that you can use to retrieve earlier communications.

_cdSeverityCode :: Maybe Text

The code for the severity level returned by the call to DescribeSeverityLevels.

_cdCaseId :: Maybe Text

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47.

_cdCcEmailAddresses :: [Text]

The email addresses that receive copies of communication about the case.

_cdDisplayId :: Maybe Text

The ID displayed for the case in the AWS Support Center. This is a numeric string.

_cdSubmittedBy :: Maybe Text

The email address of the account that submitted the case.

_cdLanguage :: Maybe Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

_cdTimeCreated :: Maybe Text

The time that the case was case created in the AWS Support Center.

_cdCategoryCode :: Maybe Text

The category of problem for the AWS Support case.

_cdServiceCode :: Maybe Text

The code for the AWS service returned by the call to DescribeServices.

data Category

A JSON-formatted name/value pair that represents the category name and category code of the problem, selected from the DescribeServices response for each AWS service.

Constructors

Category

Fields

_cyName :: Maybe Text

The category name for the support case.

_cyCode :: Maybe Text

The category code for the support case.

data Communication

A communication associated with an AWS Support case. The communication consists of the case ID, the message body, attachment information, the account email address, and the date and time of the communication.

Constructors

Communication

Fields

_cBody :: Maybe Text

The text of the communication between the customer and AWS Support.

_cCaseId :: Maybe Text

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47.

_cSubmittedBy :: Maybe Text

The email address of the account that submitted the AWS Support case.

_cTimeCreated :: Maybe Text

The time the communication was created.

_cAttachmentSet :: [AttachmentDetails]

Information about the attachments to the case communication.

data RecentCaseCommunications

The five most recent communications between you and AWS Support Center, including the IDs of any attachments to the communications. Also includes a nextToken that you can use to retrieve earlier communications.

Constructors

RecentCaseCommunications

Fields

_rccNextToken :: Maybe Text

A resumption point for pagination.

_rccCommunications :: [Communication]

The five most recent communications associated with the case.

data Service

Information about an AWS service returned by the DescribeServices operation.

Constructors

Service

Fields

_ssfCategories :: [Category]

A list of categories that describe the type of support issue a case describes. Categories consist of a category name and a category code. Category names and codes are passed to AWS Support when you call CreateCase.

_ssfName :: Maybe Text

The friendly name for an AWS service. The Code element contains the corresponding code.

_ssfCode :: Maybe Text

The code for an AWS service returned by the DescribeServices response. The Name element contains the corresponding friendly name.

data SeverityLevel

A code and name pair that represent a severity level that can be applied to a support case.

Constructors

SeverityLevel

Fields

_snName :: Maybe Text

The name of the severity level that corresponds to the severity level code.

_snCode :: Maybe Text

One of four values: "low," "medium," "high," and "urgent". These values correspond to response times returned to the caller in SeverityLevel.name.

data TrustedAdvisorCheckDescription

The description and metadata for a Trusted Advisor check.

Constructors

TrustedAdvisorCheckDescription

Fields

_tacdCategory :: Text

The category of the Trusted Advisor check.

_tacdName :: Text

The display name for the Trusted Advisor check.

_tacdMetadata :: [Text]

The column headings for the data returned by the Trusted Advisor check. The order of the headings corresponds to the order of the data in the Metadata element of the TrustedAdvisorResourceDetail for the check. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

_tacdId :: Text

The unique identifier for the Trusted Advisor check.

_tacdDescription :: Text

The description of the Trusted Advisor check, which includes the alert criteria and recommended actions (contains HTML markup).

data TrustedAdvisorCheckRefreshStatus

The current refresh status for a check, including the amount of time until the check is eligible for refresh.

Constructors

TrustedAdvisorCheckRefreshStatus

Fields

_tacrsMillisUntilNextRefreshable :: Integer

The amount of time, in milliseconds, until the Trusted Advisor check is eligible for refresh.

_tacrsStatus :: Text

The status of the Trusted Advisor check for which a refresh has been requested: "none", "enqueued", "processing", "success", or "abandoned".

_tacrsCheckId :: Text

The unique identifier for the Trusted Advisor check.

data TrustedAdvisorCheckResult

The detailed results of the Trusted Advisor check.

Constructors

TrustedAdvisorCheckResult

Fields

_tacrCategorySpecificSummary :: TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

_tacrStatus :: Text

The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

_tacrCheckId :: Text

The unique identifier for the Trusted Advisor check.

_tacrFlaggedResources :: [TrustedAdvisorResourceDetail]

The details about each resource listed in the check result.

_tacrResourcesSummary :: TrustedAdvisorResourcesSummary

Details about AWS resources that were analyzed in a call to Trusted Advisor DescribeTrustedAdvisorCheckSummaries.

_tacrTimestamp :: Text

The time of the last refresh of the check.

data TrustedAdvisorCheckSummary

A summary of a Trusted Advisor check result, including the alert status, last refresh, and number of resources examined.

Constructors

TrustedAdvisorCheckSummary

Fields

_tacsCategorySpecificSummary :: TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

_tacsStatus :: Text

The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

_tacsCheckId :: Text

The unique identifier for the Trusted Advisor check.

_tacsResourcesSummary :: TrustedAdvisorResourcesSummary

Details about AWS resources that were analyzed in a call to Trusted Advisor DescribeTrustedAdvisorCheckSummaries.

_tacsTimestamp :: Text

The time of the last refresh of the check.

_tacsHasFlaggedResources :: Maybe Bool

Specifies whether the Trusted Advisor check has flagged resources.

data TrustedAdvisorCostOptimizingSummary

The summary information about cost savings for a Trusted Advisor check that is in the Cost Optimizing category.

Constructors

TrustedAdvisorCostOptimizingSummary

Fields

_tacosEstimatedMonthlySavings :: Double

The estimated monthly savings that might be realized if the recommended actions are taken.

_tacosEstimatedPercentMonthlySavings :: Double

The estimated percentage of savings that might be realized if the recommended actions are taken.

data TrustedAdvisorResourceDetail

Contains information about a resource identified by a Trusted Advisor check.

Constructors

TrustedAdvisorResourceDetail

Fields

_tardStatus :: Text

The status code for the resource identified in the Trusted Advisor check.

_tardResourceId :: Text

The unique identifier for the identified resource.

_tardIsSuppressed :: Maybe Bool

Specifies whether the AWS resource was ignored by Trusted Advisor because it was marked as suppressed by the user.

_tardMetadata :: [Text]

Additional information about the identified resource. The exact metadata and its order can be obtained by inspecting the TrustedAdvisorCheckDescription object returned by the call to DescribeTrustedAdvisorChecks. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

_tardRegion :: Text

The AWS region in which the identified resource is located.

data TrustedAdvisorResourcesSummary

Details about AWS resources that were analyzed in a call to Trusted Advisor DescribeTrustedAdvisorCheckSummaries.

Constructors

TrustedAdvisorResourcesSummary

Fields

_tarsResourcesIgnored :: Integer

The number of AWS resources ignored by Trusted Advisor because information was unavailable.

_tarsResourcesProcessed :: Integer

The number of AWS resources that were analyzed by the Trusted Advisor check.

_tarsResourcesFlagged :: Integer

The number of AWS resources that were flagged (listed) by the Trusted Advisor check.

_tarsResourcesSuppressed :: Integer

The number of AWS resources ignored by Trusted Advisor because they were marked as suppressed by the user.

tacssCostOptimizing :: Iso' TrustedAdvisorCategorySpecificSummary (Maybe TrustedAdvisorCostOptimizingSummary)

trustedAdvisorCategorySpecificSummary :: Iso' (Maybe TrustedAdvisorCostOptimizingSummary) TrustedAdvisorCategorySpecificSummary

aFileName :: Lens' Attachment (Maybe Text)

aData :: Lens' Attachment (Maybe Base64)

adFileName :: Lens' AttachmentDetails (Maybe Text)

adAttachmentId :: Lens' AttachmentDetails (Maybe Text)

cdTimeCreated :: Lens' CaseDetails (Maybe Text)

cdSubmittedBy :: Lens' CaseDetails (Maybe Text)

cdSubject :: Lens' CaseDetails (Maybe Text)

cdStatus :: Lens' CaseDetails (Maybe Text)

cdSeverityCode :: Lens' CaseDetails (Maybe Text)

cdServiceCode :: Lens' CaseDetails (Maybe Text)

cdRecentCommunications :: Lens' CaseDetails (Maybe RecentCaseCommunications)

cdLanguage :: Lens' CaseDetails (Maybe Text)

cdDisplayId :: Lens' CaseDetails (Maybe Text)

cdCcEmailAddresses :: Lens' CaseDetails [Text]

cdCategoryCode :: Lens' CaseDetails (Maybe Text)

cdCaseId :: Lens' CaseDetails (Maybe Text)

cyName :: Lens' Category (Maybe Text)

cyCode :: Lens' Category (Maybe Text)

cTimeCreated :: Lens' Communication (Maybe Text)

cSubmittedBy :: Lens' Communication (Maybe Text)

cCaseId :: Lens' Communication (Maybe Text)

cBody :: Lens' Communication (Maybe Text)

cAttachmentSet :: Lens' Communication [AttachmentDetails]

rccNextToken :: Lens' RecentCaseCommunications (Maybe Text)

rccCommunications :: Lens' RecentCaseCommunications [Communication]

ssfName :: Lens' Service (Maybe Text)

ssfCode :: Lens' Service (Maybe Text)

ssfCategories :: Lens' Service [Category]

snName :: Lens' SeverityLevel (Maybe Text)

snCode :: Lens' SeverityLevel (Maybe Text)

tacdName :: Lens' TrustedAdvisorCheckDescription Text

tacdMetadata :: Lens' TrustedAdvisorCheckDescription [Text]

tacdId :: Lens' TrustedAdvisorCheckDescription Text

tacdDescription :: Lens' TrustedAdvisorCheckDescription Text

tacdCategory :: Lens' TrustedAdvisorCheckDescription Text

tacrsStatus :: Lens' TrustedAdvisorCheckRefreshStatus Text

tacrsMillisUntilNextRefreshable :: Lens' TrustedAdvisorCheckRefreshStatus Integer

tacrsCheckId :: Lens' TrustedAdvisorCheckRefreshStatus Text

tacrTimestamp :: Lens' TrustedAdvisorCheckResult Text

tacrStatus :: Lens' TrustedAdvisorCheckResult Text

tacrResourcesSummary :: Lens' TrustedAdvisorCheckResult TrustedAdvisorResourcesSummary

tacrFlaggedResources :: Lens' TrustedAdvisorCheckResult [TrustedAdvisorResourceDetail]

tacrCheckId :: Lens' TrustedAdvisorCheckResult Text

tacrCategorySpecificSummary :: Lens' TrustedAdvisorCheckResult TrustedAdvisorCategorySpecificSummary

tacsTimestamp :: Lens' TrustedAdvisorCheckSummary Text

tacsStatus :: Lens' TrustedAdvisorCheckSummary Text

tacsResourcesSummary :: Lens' TrustedAdvisorCheckSummary TrustedAdvisorResourcesSummary

tacsHasFlaggedResources :: Lens' TrustedAdvisorCheckSummary (Maybe Bool)

tacsCheckId :: Lens' TrustedAdvisorCheckSummary Text

tacsCategorySpecificSummary :: Lens' TrustedAdvisorCheckSummary TrustedAdvisorCategorySpecificSummary

tacosEstimatedPercentMonthlySavings :: Lens' TrustedAdvisorCostOptimizingSummary Double

tacosEstimatedMonthlySavings :: Lens' TrustedAdvisorCostOptimizingSummary Double

tardStatus :: Lens' TrustedAdvisorResourceDetail Text

tardResourceId :: Lens' TrustedAdvisorResourceDetail Text

tardRegion :: Lens' TrustedAdvisorResourceDetail Text

tardMetadata :: Lens' TrustedAdvisorResourceDetail [Text]

tardIsSuppressed :: Lens' TrustedAdvisorResourceDetail (Maybe Bool)

tarsResourcesSuppressed :: Lens' TrustedAdvisorResourcesSummary Integer

tarsResourcesProcessed :: Lens' TrustedAdvisorResourcesSummary Integer

tarsResourcesIgnored :: Lens' TrustedAdvisorResourcesSummary Integer

tarsResourcesFlagged :: Lens' TrustedAdvisorResourcesSummary Integer