Network.AWS.Support.V2013_04_15.DescribeServices

Description

Returns the current list of AWS services and a list of service categories that applies to each one. You then use service names and categories in your CreateCase requests. Each AWS service has its own set of categories. The service codes and category codes correspond to the values that are displayed in the Service and Category drop-down lists on the AWS Support Center Open a new case page. The values in those fields, however, do not necessarily match the service codes and categories returned by the DescribeServices request. Always use the service codes and categories obtained programmatically. This practice ensures that you always have the most recent set of service and category codes.

Documentation

describeServices :: DescribeServices

Minimum specification for a DescribeServices request.

data DescribeServices

Constructors

DescribeServices

Fields

_dsrLanguage :: Maybe Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

_dsrServiceCodeList :: [Text]

A JSON-formatted list of service codes available for AWS services.

dsrServiceCodeList :: Lens' DescribeServices [Text]

dsrLanguage :: Lens' DescribeServices (Maybe Text)

data DescribeServicesResponse

Constructors

DescribeServicesResponse

Fields

_dssServices :: [Service]

A JSON-formatted list of AWS services.

dssServices :: Iso' DescribeServicesResponse [Service]