Network.AWS.Synthetics.GetCanary

Description

Retrieves complete information about one canary. You must specify the name of the canary that you want. To get a list of canaries and their names, use DescribeCanaries.

Creating a Request

data GetCanary

See: newGetCanary smart constructor.

Constructors

GetCanary'

Fields

name :: Text

The name of the canary that you want details for.

newGetCanary

Arguments

:: Text
$sel:name:GetCanary'

-> GetCanary

Create a value of GetCanary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetCanary', getCanary_name - The name of the canary that you want details for.

Request Lenses

getCanary_name :: Lens' GetCanary Text

The name of the canary that you want details for.

Destructuring the Response

data GetCanaryResponse

See: newGetCanaryResponse smart constructor.

Constructors

GetCanaryResponse'

Fields

canary :: Maybe Canary

A strucure that contains the full information about the canary.

httpStatus :: Int

The response's http status code.

newGetCanaryResponse

Arguments

:: Int
$sel:httpStatus:GetCanaryResponse'

-> GetCanaryResponse

Create a value of GetCanaryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:canary:GetCanaryResponse', getCanaryResponse_canary - A strucure that contains the full information about the canary.

$sel:httpStatus:GetCanaryResponse', getCanaryResponse_httpStatus - The response's http status code.

Response Lenses

getCanaryResponse_canary :: Lens' GetCanaryResponse (Maybe Canary)

A strucure that contains the full information about the canary.

getCanaryResponse_httpStatus :: Lens' GetCanaryResponse Int

The response's http status code.