Network.AWS.Synthetics.Lens

Description

Operations

UpdateCanary

updateCanary_successRetentionPeriodInDays :: Lens' UpdateCanary (Maybe Natural)

The number of days to retain data about successful runs of this canary.

updateCanary_schedule :: Lens' UpdateCanary (Maybe CanaryScheduleInput)

A structure that contains information about how often the canary is to run, and when these runs are to stop.

updateCanary_artifactS3Location :: Lens' UpdateCanary (Maybe Text)

The location in Amazon S3 where Synthetics stores artifacts from the test runs of this canary. Artifacts include the log file, screenshots, and HAR files. The name of the S3 bucket can't include a period (.).

updateCanary_runConfig :: Lens' UpdateCanary (Maybe CanaryRunConfigInput)

A structure that contains the timeout value that is used for each individual run of the canary.

updateCanary_executionRoleArn :: Lens' UpdateCanary (Maybe Text)

The ARN of the IAM role to be used to run the canary. This role must already exist, and must include lambda.amazonaws.com as a principal in the trust policy. The role must also have the following permissions:

s3:PutObject

s3:GetBucketLocation

s3:ListAllMyBuckets

cloudwatch:PutMetricData

logs:CreateLogGroup

logs:CreateLogStream

logs:CreateLogStream

updateCanary_runtimeVersion :: Lens' UpdateCanary (Maybe Text)

Specifies the runtime version to use for the canary. For a list of valid runtime versions and for more information about runtime versions, see Canary Runtime Versions.

updateCanary_failureRetentionPeriodInDays :: Lens' UpdateCanary (Maybe Natural)

The number of days to retain data about failed runs of this canary.

updateCanary_artifactConfig :: Lens' UpdateCanary (Maybe ArtifactConfigInput)

A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

updateCanary_vpcConfig :: Lens' UpdateCanary (Maybe VpcConfigInput)

If this canary is to test an endpoint in a VPC, this structure contains information about the subnet and security groups of the VPC endpoint. For more information, see Running a Canary in a VPC.

updateCanary_visualReference :: Lens' UpdateCanary (Maybe VisualReferenceInput)

Defines the screenshots to use as the baseline for comparisons during visual monitoring comparisons during future runs of this canary. If you omit this parameter, no changes are made to any baseline screenshots that the canary might be using already.

Visual monitoring is supported only on canaries running the syn-puppeteer-node-3.2 runtime or later. For more information, see Visual monitoring and Visual monitoring blueprint

updateCanary_code :: Lens' UpdateCanary (Maybe CanaryCodeInput)

A structure that includes the entry point from which the canary should start running your script. If the script is stored in an S3 bucket, the bucket name, key, and version are also included.

updateCanary_name :: Lens' UpdateCanary Text

The name of the canary that you want to update. To find the names of your canaries, use DescribeCanaries.

You cannot change the name of a canary that has already been created.

updateCanaryResponse_httpStatus :: Lens' UpdateCanaryResponse Int

The response's http status code.

DeleteCanary

deleteCanary_name :: Lens' DeleteCanary Text

The name of the canary that you want to delete. To find the names of your canaries, use DescribeCanaries.

deleteCanaryResponse_httpStatus :: Lens' DeleteCanaryResponse Int

The response's http status code.

CreateCanary

createCanary_successRetentionPeriodInDays :: Lens' CreateCanary (Maybe Natural)

The number of days to retain data about successful runs of this canary. If you omit this field, the default of 31 days is used. The valid range is 1 to 455 days.

createCanary_runConfig :: Lens' CreateCanary (Maybe CanaryRunConfigInput)

A structure that contains the configuration for individual canary runs, such as timeout value.

createCanary_failureRetentionPeriodInDays :: Lens' CreateCanary (Maybe Natural)

The number of days to retain data about failed runs of this canary. If you omit this field, the default of 31 days is used. The valid range is 1 to 455 days.

createCanary_artifactConfig :: Lens' CreateCanary (Maybe ArtifactConfigInput)

A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

createCanary_vpcConfig :: Lens' CreateCanary (Maybe VpcConfigInput)

If this canary is to test an endpoint in a VPC, this structure contains information about the subnet and security groups of the VPC endpoint. For more information, see Running a Canary in a VPC.

createCanary_tags :: Lens' CreateCanary (Maybe (HashMap Text Text))

A list of key-value pairs to associate with the canary. You can associate as many as 50 tags with a canary.

Tags can help you organize and categorize your resources. You can also use them to scope user permissions, by granting a user permission to access or change only the resources that have certain tag values.

createCanary_name :: Lens' CreateCanary Text

The name for this canary. Be sure to give it a descriptive name that distinguishes it from other canaries in your account.

Do not include secrets or proprietary information in your canary names. The canary name makes up part of the canary ARN, and the ARN is included in outbound calls over the internet. For more information, see Security Considerations for Synthetics Canaries.

createCanary_code :: Lens' CreateCanary CanaryCodeInput

A structure that includes the entry point from which the canary should start running your script. If the script is stored in an S3 bucket, the bucket name, key, and version are also included.

createCanary_artifactS3Location :: Lens' CreateCanary Text

The location in Amazon S3 where Synthetics stores artifacts from the test runs of this canary. Artifacts include the log file, screenshots, and HAR files. The name of the S3 bucket can't include a period (.).

createCanary_executionRoleArn :: Lens' CreateCanary Text

The ARN of the IAM role to be used to run the canary. This role must already exist, and must include lambda.amazonaws.com as a principal in the trust policy. The role must also have the following permissions:

s3:PutObject

s3:GetBucketLocation

s3:ListAllMyBuckets

cloudwatch:PutMetricData

logs:CreateLogGroup

logs:CreateLogStream

logs:PutLogEvents

createCanary_schedule :: Lens' CreateCanary CanaryScheduleInput

A structure that contains information about how often the canary is to run and when these test runs are to stop.

createCanary_runtimeVersion :: Lens' CreateCanary Text

Specifies the runtime version to use for the canary. For a list of valid runtime versions and more information about runtime versions, see Canary Runtime Versions.

createCanaryResponse_canary :: Lens' CreateCanaryResponse (Maybe Canary)

The full details about the canary you have created.

createCanaryResponse_httpStatus :: Lens' CreateCanaryResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The ARN of the canary that you want to view tags for.

The ARN format of a canary is arn:aws:synthetics:Region:account-id:canary:canary-name .

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The list of tag keys and values associated with the canary that you specified.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

GetCanaryRuns

getCanaryRuns_nextToken :: Lens' GetCanaryRuns (Maybe Text)

A token that indicates that there is more data available. You can use this token in a subsequent GetCanaryRuns operation to retrieve the next set of results.

getCanaryRuns_maxResults :: Lens' GetCanaryRuns (Maybe Natural)

Specify this parameter to limit how many runs are returned each time you use the GetCanaryRuns operation. If you omit this parameter, the default of 100 is used.

getCanaryRuns_name :: Lens' GetCanaryRuns Text

The name of the canary that you want to see runs for.

getCanaryRunsResponse_nextToken :: Lens' GetCanaryRunsResponse (Maybe Text)

A token that indicates that there is more data available. You can use this token in a subsequent GetCanaryRuns operation to retrieve the next set of results.

getCanaryRunsResponse_canaryRuns :: Lens' GetCanaryRunsResponse (Maybe [CanaryRun])

An array of structures. Each structure contains the details of one of the retrieved canary runs.

getCanaryRunsResponse_httpStatus :: Lens' GetCanaryRunsResponse Int

The response's http status code.

GetCanary

getCanary_name :: Lens' GetCanary Text

The name of the canary that you want details for.

getCanaryResponse_canary :: Lens' GetCanaryResponse (Maybe Canary)

A strucure that contains the full information about the canary.

getCanaryResponse_httpStatus :: Lens' GetCanaryResponse Int

The response's http status code.

DescribeRuntimeVersions

describeRuntimeVersions_nextToken :: Lens' DescribeRuntimeVersions (Maybe Text)

A token that indicates that there is more data available. You can use this token in a subsequent DescribeRuntimeVersions operation to retrieve the next set of results.

describeRuntimeVersions_maxResults :: Lens' DescribeRuntimeVersions (Maybe Natural)

Specify this parameter to limit how many runs are returned each time you use the DescribeRuntimeVersions operation. If you omit this parameter, the default of 100 is used.

describeRuntimeVersionsResponse_runtimeVersions :: Lens' DescribeRuntimeVersionsResponse (Maybe [RuntimeVersion])

An array of objects that display the details about each Synthetics canary runtime version.

describeRuntimeVersionsResponse_nextToken :: Lens' DescribeRuntimeVersionsResponse (Maybe Text)

A token that indicates that there is more data available. You can use this token in a subsequent DescribeRuntimeVersions operation to retrieve the next set of results.

describeRuntimeVersionsResponse_httpStatus :: Lens' DescribeRuntimeVersionsResponse Int

The response's http status code.

DescribeCanariesLastRun

describeCanariesLastRun_nextToken :: Lens' DescribeCanariesLastRun (Maybe Text)

A token that indicates that there is more data available. You can use this token in a subsequent DescribeCanaries operation to retrieve the next set of results.

describeCanariesLastRun_maxResults :: Lens' DescribeCanariesLastRun (Maybe Natural)

Specify this parameter to limit how many runs are returned each time you use the DescribeLastRun operation. If you omit this parameter, the default of 100 is used.

describeCanariesLastRunResponse_nextToken :: Lens' DescribeCanariesLastRunResponse (Maybe Text)

A token that indicates that there is more data available. You can use this token in a subsequent DescribeCanariesLastRun operation to retrieve the next set of results.

describeCanariesLastRunResponse_canariesLastRun :: Lens' DescribeCanariesLastRunResponse (Maybe [CanaryLastRun])

An array that contains the information from the most recent run of each canary.

describeCanariesLastRunResponse_httpStatus :: Lens' DescribeCanariesLastRunResponse Int

The response's http status code.

StartCanary

startCanary_name :: Lens' StartCanary Text

The name of the canary that you want to run. To find canary names, use DescribeCanaries.

startCanaryResponse_httpStatus :: Lens' StartCanaryResponse Int

The response's http status code.

DescribeCanaries

describeCanaries_nextToken :: Lens' DescribeCanaries (Maybe Text)

A token that indicates that there is more data available. You can use this token in a subsequent operation to retrieve the next set of results.

describeCanaries_maxResults :: Lens' DescribeCanaries (Maybe Natural)

Specify this parameter to limit how many canaries are returned each time you use the DescribeCanaries operation. If you omit this parameter, the default of 100 is used.

describeCanariesResponse_canaries :: Lens' DescribeCanariesResponse (Maybe [Canary])

Returns an array. Each item in the array contains the full information about one canary.

describeCanariesResponse_nextToken :: Lens' DescribeCanariesResponse (Maybe Text)

A token that indicates that there is more data available. You can use this token in a subsequent DescribeCanaries operation to retrieve the next set of results.

describeCanariesResponse_httpStatus :: Lens' DescribeCanariesResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The ARN of the canary that you're adding tags to.

The ARN format of a canary is arn:aws:synthetics:Region:account-id:canary:canary-name .

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The list of key-value pairs to associate with the canary.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The ARN of the canary that you're removing tags from.

The ARN format of a canary is arn:aws:synthetics:Region:account-id:canary:canary-name .

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

The list of tag keys to remove from the resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

StopCanary

stopCanary_name :: Lens' StopCanary Text

The name of the canary that you want to stop. To find the names of your canaries, use DescribeCanaries.

stopCanaryResponse_httpStatus :: Lens' StopCanaryResponse Int

The response's http status code.

Types

ArtifactConfigInput

artifactConfigInput_s3Encryption :: Lens' ArtifactConfigInput (Maybe S3EncryptionConfig)

A structure that contains the configuration of the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3. Artifact encryption functionality is available only for canaries that use Synthetics runtime version syn-nodejs-puppeteer-3.3 or later. For more information, see Encrypting canary artifacts

ArtifactConfigOutput

artifactConfigOutput_s3Encryption :: Lens' ArtifactConfigOutput (Maybe S3EncryptionConfig)

A structure that contains the configuration of encryption settings for canary artifacts that are stored in Amazon S3.

BaseScreenshot

baseScreenshot_ignoreCoordinates :: Lens' BaseScreenshot (Maybe [Text])

Coordinates that define the part of a screen to ignore during screenshot comparisons. To obtain the coordinates to use here, use the CloudWatch Logs console to draw the boundaries on the screen. For more information, see {LINK}

baseScreenshot_screenshotName :: Lens' BaseScreenshot Text

The name of the screenshot. This is generated the first time the canary is run after the UpdateCanary operation that specified for this canary to perform visual monitoring.

Canary

canary_status :: Lens' Canary (Maybe CanaryStatus)

A structure that contains information about the canary's status.

canary_successRetentionPeriodInDays :: Lens' Canary (Maybe Natural)

The number of days to retain data about successful runs of this canary.

canary_schedule :: Lens' Canary (Maybe CanaryScheduleOutput)

A structure that contains information about how often the canary is to run, and when these runs are to stop.

canary_artifactS3Location :: Lens' Canary (Maybe Text)

The location in Amazon S3 where Synthetics stores artifacts from the runs of this canary. Artifacts include the log file, screenshots, and HAR files.

canary_runConfig :: Lens' Canary (Maybe CanaryRunConfigOutput)

Undocumented member.

canary_executionRoleArn :: Lens' Canary (Maybe Text)

The ARN of the IAM role used to run the canary. This role must include lambda.amazonaws.com as a principal in the trust policy.

canary_runtimeVersion :: Lens' Canary (Maybe Text)

Specifies the runtime version to use for the canary. For more information about runtime versions, see Canary Runtime Versions.

canary_failureRetentionPeriodInDays :: Lens' Canary (Maybe Natural)

The number of days to retain data about failed runs of this canary.

canary_artifactConfig :: Lens' Canary (Maybe ArtifactConfigOutput)

A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

canary_vpcConfig :: Lens' Canary (Maybe VpcConfigOutput)

Undocumented member.

canary_visualReference :: Lens' Canary (Maybe VisualReferenceOutput)

If this canary performs visual monitoring by comparing screenshots, this structure contains the ID of the canary run to use as the baseline for screenshots, and the coordinates of any parts of the screen to ignore during the visual monitoring comparison.

canary_name :: Lens' Canary (Maybe Text)

The name of the canary.

canary_id :: Lens' Canary (Maybe Text)

The unique ID of this canary.

canary_code :: Lens' Canary (Maybe CanaryCodeOutput)

Undocumented member.

canary_timeline :: Lens' Canary (Maybe CanaryTimeline)

A structure that contains information about when the canary was created, modified, and most recently run.

canary_engineArn :: Lens' Canary (Maybe Text)

The ARN of the Lambda function that is used as your canary's engine. For more information about Lambda ARN format, see Resources and Conditions for Lambda Actions.

canary_tags :: Lens' Canary (Maybe (HashMap Text Text))

The list of key-value pairs that are associated with the canary.

CanaryCodeInput

canaryCodeInput_s3Key :: Lens' CanaryCodeInput (Maybe Text)

The S3 key of your script. For more information, see Working with Amazon S3 Objects.

canaryCodeInput_s3Version :: Lens' CanaryCodeInput (Maybe Text)

The S3 version ID of your script.

canaryCodeInput_zipFile :: Lens' CanaryCodeInput (Maybe ByteString)

If you input your canary script directly into the canary instead of referring to an S3 location, the value of this parameter is the base64-encoded contents of the .zip file that contains the script. It must be smaller than 256 Kb.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

canaryCodeInput_s3Bucket :: Lens' CanaryCodeInput (Maybe Text)

If your canary script is located in S3, specify the bucket name here. Do not include s3:// as the start of the bucket name.

canaryCodeInput_handler :: Lens' CanaryCodeInput Text

The entry point to use for the source code when running the canary. This value must end with the string .handler. The string is limited to 29 characters or fewer.

CanaryCodeOutput

canaryCodeOutput_sourceLocationArn :: Lens' CanaryCodeOutput (Maybe Text)

The ARN of the Lambda layer where Synthetics stores the canary script code.

canaryCodeOutput_handler :: Lens' CanaryCodeOutput (Maybe Text)

The entry point to use for the source code when running the canary.

CanaryLastRun

canaryLastRun_canaryName :: Lens' CanaryLastRun (Maybe Text)

The name of the canary.

canaryLastRun_lastRun :: Lens' CanaryLastRun (Maybe CanaryRun)

The results from this canary's most recent run.

CanaryRun

canaryRun_status :: Lens' CanaryRun (Maybe CanaryRunStatus)

The status of this run.

canaryRun_artifactS3Location :: Lens' CanaryRun (Maybe Text)

The location where the canary stored artifacts from the run. Artifacts include the log file, screenshots, and HAR files.

canaryRun_name :: Lens' CanaryRun (Maybe Text)

The name of the canary.

canaryRun_id :: Lens' CanaryRun (Maybe Text)

A unique ID that identifies this canary run.

canaryRun_timeline :: Lens' CanaryRun (Maybe CanaryRunTimeline)

A structure that contains the start and end times of this run.

CanaryRunConfigInput

canaryRunConfigInput_timeoutInSeconds :: Lens' CanaryRunConfigInput (Maybe Natural)

How long the canary is allowed to run before it must stop. You can't set this time to be longer than the frequency of the runs of this canary.

If you omit this field, the frequency of the canary is used as this value, up to a maximum of 14 minutes.

canaryRunConfigInput_environmentVariables :: Lens' CanaryRunConfigInput (Maybe (HashMap Text Text))

Specifies the keys and values to use for any environment variables used in the canary script. Use the following format:

{ "key1" : "value1", "key2" : "value2", ...}

Keys must start with a letter and be at least two characters. The total size of your environment variables cannot exceed 4 KB. You can't specify any Lambda reserved environment variables as the keys for your environment variables. For more information about reserved keys, see Runtime environment variables.

canaryRunConfigInput_activeTracing :: Lens' CanaryRunConfigInput (Maybe Bool)

Specifies whether this canary is to use active X-Ray tracing when it runs. Active tracing enables this canary run to be displayed in the ServiceLens and X-Ray service maps even if the canary does not hit an endpoint that has X-Ray tracing enabled. Using X-Ray tracing incurs charges. For more information, see Canaries and X-Ray tracing.

You can enable active tracing only for canaries that use version syn-nodejs-2.0 or later for their canary runtime.

canaryRunConfigInput_memoryInMB :: Lens' CanaryRunConfigInput (Maybe Natural)

The maximum amount of memory available to the canary while it is running, in MB. This value must be a multiple of 64.

CanaryRunConfigOutput

canaryRunConfigOutput_timeoutInSeconds :: Lens' CanaryRunConfigOutput (Maybe Natural)

How long the canary is allowed to run before it must stop.

canaryRunConfigOutput_activeTracing :: Lens' CanaryRunConfigOutput (Maybe Bool)

Displays whether this canary run used active X-Ray tracing.

canaryRunConfigOutput_memoryInMB :: Lens' CanaryRunConfigOutput (Maybe Natural)

The maximum amount of memory available to the canary while it is running, in MB. This value must be a multiple of 64.

CanaryRunStatus

canaryRunStatus_state :: Lens' CanaryRunStatus (Maybe CanaryRunState)

The current state of the run.

canaryRunStatus_stateReason :: Lens' CanaryRunStatus (Maybe Text)

If run of the canary failed, this field contains the reason for the error.

canaryRunStatus_stateReasonCode :: Lens' CanaryRunStatus (Maybe CanaryRunStateReasonCode)

If this value is CANARY_FAILURE, an exception occurred in the canary code. If this value is EXECUTION_FAILURE, an exception occurred in CloudWatch Synthetics.

CanaryRunTimeline

canaryRunTimeline_completed :: Lens' CanaryRunTimeline (Maybe UTCTime)

The end time of the run.

canaryRunTimeline_started :: Lens' CanaryRunTimeline (Maybe UTCTime)

The start time of the run.

CanaryScheduleInput

canaryScheduleInput_durationInSeconds :: Lens' CanaryScheduleInput (Maybe Natural)

How long, in seconds, for the canary to continue making regular runs according to the schedule in the Expression value. If you specify 0, the canary continues making runs until you stop it. If you omit this field, the default of 0 is used.

canaryScheduleInput_expression :: Lens' CanaryScheduleInput Text

A rate expression or a cron expression that defines how often the canary is to run.

For a rate expression, The syntax is rate(number unit). unit can be minute, minutes, or hour.

For example, rate(1 minute) runs the canary once a minute, rate(10 minutes) runs it once every 10 minutes, and rate(1 hour) runs it once every hour. You can specify a frequency between rate(1 minute) and rate(1 hour).

Specifying rate(0 minute) or rate(0 hour) is a special value that causes the canary to run only once when it is started.

Use cron(expression) to specify a cron expression. You can't schedule a canary to wait for more than a year before running. For information about the syntax for cron expressions, see Scheduling canary runs using cron.

CanaryScheduleOutput

canaryScheduleOutput_durationInSeconds :: Lens' CanaryScheduleOutput (Maybe Natural)

How long, in seconds, for the canary to continue making regular runs after it was created. The runs are performed according to the schedule in the Expression value.

canaryScheduleOutput_expression :: Lens' CanaryScheduleOutput (Maybe Text)

A rate expression or a cron expression that defines how often the canary is to run.

For a rate expression, The syntax is rate(number unit). unit can be minute, minutes, or hour.

For example, rate(1 minute) runs the canary once a minute, rate(10 minutes) runs it once every 10 minutes, and rate(1 hour) runs it once every hour. You can specify a frequency between rate(1 minute) and rate(1 hour).

Specifying rate(0 minute) or rate(0 hour) is a special value that causes the canary to run only once when it is started.

Use cron(expression) to specify a cron expression. For information about the syntax for cron expressions, see Scheduling canary runs using cron.

CanaryStatus

canaryStatus_state :: Lens' CanaryStatus (Maybe CanaryState)

The current state of the canary.

canaryStatus_stateReason :: Lens' CanaryStatus (Maybe Text)

If the canary has insufficient permissions to run, this field provides more details.

canaryStatus_stateReasonCode :: Lens' CanaryStatus (Maybe CanaryStateReasonCode)

If the canary cannot run or has failed, this field displays the reason.

CanaryTimeline

canaryTimeline_created :: Lens' CanaryTimeline (Maybe UTCTime)

The date and time the canary was created.

canaryTimeline_lastStarted :: Lens' CanaryTimeline (Maybe UTCTime)

The date and time that the canary's most recent run started.

canaryTimeline_lastStopped :: Lens' CanaryTimeline (Maybe UTCTime)

The date and time that the canary's most recent run ended.

canaryTimeline_lastModified :: Lens' CanaryTimeline (Maybe UTCTime)

The date and time the canary was most recently modified.

RuntimeVersion

runtimeVersion_versionName :: Lens' RuntimeVersion (Maybe Text)

The name of the runtime version. For a list of valid runtime versions, see Canary Runtime Versions.

runtimeVersion_deprecationDate :: Lens' RuntimeVersion (Maybe UTCTime)

If this runtime version is deprecated, this value is the date of deprecation.

runtimeVersion_releaseDate :: Lens' RuntimeVersion (Maybe UTCTime)

The date that the runtime version was released.

runtimeVersion_description :: Lens' RuntimeVersion (Maybe Text)

A description of the runtime version, created by Amazon.

S3EncryptionConfig

s3EncryptionConfig_kmsKeyArn :: Lens' S3EncryptionConfig (Maybe Text)

The ARN of the customer-managed KMS key to use, if you specify SSE-KMS for EncryptionMode

s3EncryptionConfig_encryptionMode :: Lens' S3EncryptionConfig (Maybe EncryptionMode)

The encryption method to use for artifacts created by this canary. Specify SSE_S3 to use server-side encryption (SSE) with an Amazon S3-managed key. Specify SSE-KMS to use server-side encryption with a customer-managed KMS key.

If you omit this parameter, an Amazon Web Services-managed KMS key is used.

VisualReferenceInput

visualReferenceInput_baseScreenshots :: Lens' VisualReferenceInput (Maybe [BaseScreenshot])

An array of screenshots that will be used as the baseline for visual monitoring in future runs of this canary. If there is a screenshot that you don't want to be used for visual monitoring, remove it from this array.

visualReferenceInput_baseCanaryRunId :: Lens' VisualReferenceInput Text

Specifies which canary run to use the screenshots from as the baseline for future visual monitoring with this canary. Valid values are nextrun to use the screenshots from the next run after this update is made, lastrun to use the screenshots from the most recent run before this update was made, or the value of Id in the CanaryRun from any past run of this canary.

VisualReferenceOutput

visualReferenceOutput_baseScreenshots :: Lens' VisualReferenceOutput (Maybe [BaseScreenshot])

An array of screenshots that are used as the baseline for comparisons during visual monitoring.

visualReferenceOutput_baseCanaryRunId :: Lens' VisualReferenceOutput (Maybe Text)

The ID of the canary run that produced the screenshots that are used as the baseline for visual monitoring comparisons during future runs of this canary.

VpcConfigInput

vpcConfigInput_securityGroupIds :: Lens' VpcConfigInput (Maybe [Text])

The IDs of the security groups for this canary.

vpcConfigInput_subnetIds :: Lens' VpcConfigInput (Maybe [Text])

The IDs of the subnets where this canary is to run.

VpcConfigOutput

vpcConfigOutput_securityGroupIds :: Lens' VpcConfigOutput (Maybe [Text])

The IDs of the security groups for this canary.

vpcConfigOutput_subnetIds :: Lens' VpcConfigOutput (Maybe [Text])

The IDs of the subnets where this canary is to run.

vpcConfigOutput_vpcId :: Lens' VpcConfigOutput (Maybe Text)

The IDs of the VPC where this canary is to run.