Network.AWS.Synthetics.Types.CanaryTimeline

Description

Documentation

data CanaryTimeline

This structure contains information about when the canary was created and modified.

See: newCanaryTimeline smart constructor.

Constructors

CanaryTimeline'

Fields

created :: Maybe POSIX

The date and time the canary was created.

lastStarted :: Maybe POSIX

The date and time that the canary's most recent run started.

lastStopped :: Maybe POSIX

The date and time that the canary's most recent run ended.

lastModified :: Maybe POSIX

The date and time the canary was most recently modified.

newCanaryTimeline :: CanaryTimeline

Create a value of CanaryTimeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:created:CanaryTimeline', canaryTimeline_created - The date and time the canary was created.

$sel:lastStarted:CanaryTimeline', canaryTimeline_lastStarted - The date and time that the canary's most recent run started.

$sel:lastStopped:CanaryTimeline', canaryTimeline_lastStopped - The date and time that the canary's most recent run ended.

$sel:lastModified:CanaryTimeline', canaryTimeline_lastModified - The date and time the canary was most recently modified.

canaryTimeline_created :: Lens' CanaryTimeline (Maybe UTCTime)

The date and time the canary was created.

canaryTimeline_lastStarted :: Lens' CanaryTimeline (Maybe UTCTime)

The date and time that the canary's most recent run started.

canaryTimeline_lastStopped :: Lens' CanaryTimeline (Maybe UTCTime)

The date and time that the canary's most recent run ended.

canaryTimeline_lastModified :: Lens' CanaryTimeline (Maybe UTCTime)

The date and time the canary was most recently modified.