Network.AWS.Synthetics.Types.VisualReferenceOutput

Description

Documentation

data VisualReferenceOutput

If this canary performs visual monitoring by comparing screenshots, this structure contains the ID of the canary run that is used as the baseline for screenshots, and the coordinates of any parts of those screenshots that are ignored during visual monitoring comparison.

Visual monitoring is supported only on canaries running the syn-puppeteer-node-3.2 runtime or later.

See: newVisualReferenceOutput smart constructor.

Constructors

VisualReferenceOutput'

Fields

baseScreenshots :: Maybe [BaseScreenshot]

An array of screenshots that are used as the baseline for comparisons during visual monitoring.

baseCanaryRunId :: Maybe Text

The ID of the canary run that produced the screenshots that are used as the baseline for visual monitoring comparisons during future runs of this canary.

newVisualReferenceOutput :: VisualReferenceOutput

Create a value of VisualReferenceOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baseScreenshots:VisualReferenceOutput', visualReferenceOutput_baseScreenshots - An array of screenshots that are used as the baseline for comparisons during visual monitoring.

$sel:baseCanaryRunId:VisualReferenceOutput', visualReferenceOutput_baseCanaryRunId - The ID of the canary run that produced the screenshots that are used as the baseline for visual monitoring comparisons during future runs of this canary.

visualReferenceOutput_baseScreenshots :: Lens' VisualReferenceOutput (Maybe [BaseScreenshot])

An array of screenshots that are used as the baseline for comparisons during visual monitoring.

visualReferenceOutput_baseCanaryRunId :: Lens' VisualReferenceOutput (Maybe Text)

The ID of the canary run that produced the screenshots that are used as the baseline for visual monitoring comparisons during future runs of this canary.