Network.AWS.Synthetics.UpdateCanary

Description

Use this operation to change the settings of a canary that has already been created.

You can't use this operation to update the tags of an existing canary. To change the tags of an existing canary, use TagResource.

Creating a Request

data UpdateCanary

See: newUpdateCanary smart constructor.

Constructors

UpdateCanary'

Fields

successRetentionPeriodInDays :: Maybe Natural

The number of days to retain data about successful runs of this canary.

schedule :: Maybe CanaryScheduleInput

A structure that contains information about how often the canary is to run, and when these runs are to stop.

artifactS3Location :: Maybe Text

The location in Amazon S3 where Synthetics stores artifacts from the test runs of this canary. Artifacts include the log file, screenshots, and HAR files. The name of the S3 bucket can't include a period (.).

runConfig :: Maybe CanaryRunConfigInput

A structure that contains the timeout value that is used for each individual run of the canary.

executionRoleArn :: Maybe Text

The ARN of the IAM role to be used to run the canary. This role must already exist, and must include lambda.amazonaws.com as a principal in the trust policy. The role must also have the following permissions:

s3:PutObject

s3:GetBucketLocation

s3:ListAllMyBuckets

cloudwatch:PutMetricData

logs:CreateLogGroup

logs:CreateLogStream

logs:CreateLogStream

runtimeVersion :: Maybe Text

Specifies the runtime version to use for the canary. For a list of valid runtime versions and for more information about runtime versions, see Canary Runtime Versions.

failureRetentionPeriodInDays :: Maybe Natural

The number of days to retain data about failed runs of this canary.

artifactConfig :: Maybe ArtifactConfigInput

A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

vpcConfig :: Maybe VpcConfigInput

If this canary is to test an endpoint in a VPC, this structure contains information about the subnet and security groups of the VPC endpoint. For more information, see Running a Canary in a VPC.

visualReference :: Maybe VisualReferenceInput

Defines the screenshots to use as the baseline for comparisons during visual monitoring comparisons during future runs of this canary. If you omit this parameter, no changes are made to any baseline screenshots that the canary might be using already.

Visual monitoring is supported only on canaries running the syn-puppeteer-node-3.2 runtime or later. For more information, see Visual monitoring and Visual monitoring blueprint

code :: Maybe CanaryCodeInput

A structure that includes the entry point from which the canary should start running your script. If the script is stored in an S3 bucket, the bucket name, key, and version are also included.

name :: Text

The name of the canary that you want to update. To find the names of your canaries, use DescribeCanaries.

You cannot change the name of a canary that has already been created.

newUpdateCanary

Arguments

:: Text
$sel:name:UpdateCanary'

-> UpdateCanary

Create a value of UpdateCanary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:successRetentionPeriodInDays:UpdateCanary', updateCanary_successRetentionPeriodInDays - The number of days to retain data about successful runs of this canary.

$sel:schedule:UpdateCanary', updateCanary_schedule - A structure that contains information about how often the canary is to run, and when these runs are to stop.

$sel:artifactS3Location:UpdateCanary', updateCanary_artifactS3Location - The location in Amazon S3 where Synthetics stores artifacts from the test runs of this canary. Artifacts include the log file, screenshots, and HAR files. The name of the S3 bucket can't include a period (.).

$sel:runConfig:UpdateCanary', updateCanary_runConfig - A structure that contains the timeout value that is used for each individual run of the canary.

$sel:executionRoleArn:UpdateCanary', updateCanary_executionRoleArn - The ARN of the IAM role to be used to run the canary. This role must already exist, and must include lambda.amazonaws.com as a principal in the trust policy. The role must also have the following permissions:

s3:PutObject

s3:GetBucketLocation

s3:ListAllMyBuckets

cloudwatch:PutMetricData

logs:CreateLogGroup

logs:CreateLogStream

logs:CreateLogStream

$sel:runtimeVersion:UpdateCanary', updateCanary_runtimeVersion - Specifies the runtime version to use for the canary. For a list of valid runtime versions and for more information about runtime versions, see Canary Runtime Versions.

$sel:failureRetentionPeriodInDays:UpdateCanary', updateCanary_failureRetentionPeriodInDays - The number of days to retain data about failed runs of this canary.

$sel:artifactConfig:UpdateCanary', updateCanary_artifactConfig - A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

$sel:vpcConfig:UpdateCanary', updateCanary_vpcConfig - If this canary is to test an endpoint in a VPC, this structure contains information about the subnet and security groups of the VPC endpoint. For more information, see Running a Canary in a VPC.

$sel:visualReference:UpdateCanary', updateCanary_visualReference - Defines the screenshots to use as the baseline for comparisons during visual monitoring comparisons during future runs of this canary. If you omit this parameter, no changes are made to any baseline screenshots that the canary might be using already.

Visual monitoring is supported only on canaries running the syn-puppeteer-node-3.2 runtime or later. For more information, see Visual monitoring and Visual monitoring blueprint

$sel:code:UpdateCanary', updateCanary_code - A structure that includes the entry point from which the canary should start running your script. If the script is stored in an S3 bucket, the bucket name, key, and version are also included.

$sel:name:UpdateCanary', updateCanary_name - The name of the canary that you want to update. To find the names of your canaries, use DescribeCanaries.

You cannot change the name of a canary that has already been created.

Request Lenses

updateCanary_successRetentionPeriodInDays :: Lens' UpdateCanary (Maybe Natural)

The number of days to retain data about successful runs of this canary.

updateCanary_schedule :: Lens' UpdateCanary (Maybe CanaryScheduleInput)

A structure that contains information about how often the canary is to run, and when these runs are to stop.

updateCanary_artifactS3Location :: Lens' UpdateCanary (Maybe Text)

The location in Amazon S3 where Synthetics stores artifacts from the test runs of this canary. Artifacts include the log file, screenshots, and HAR files. The name of the S3 bucket can't include a period (.).

updateCanary_runConfig :: Lens' UpdateCanary (Maybe CanaryRunConfigInput)

A structure that contains the timeout value that is used for each individual run of the canary.

updateCanary_executionRoleArn :: Lens' UpdateCanary (Maybe Text)

The ARN of the IAM role to be used to run the canary. This role must already exist, and must include lambda.amazonaws.com as a principal in the trust policy. The role must also have the following permissions:

s3:PutObject

s3:GetBucketLocation

s3:ListAllMyBuckets

cloudwatch:PutMetricData

logs:CreateLogGroup

logs:CreateLogStream

logs:CreateLogStream

updateCanary_runtimeVersion :: Lens' UpdateCanary (Maybe Text)

Specifies the runtime version to use for the canary. For a list of valid runtime versions and for more information about runtime versions, see Canary Runtime Versions.

updateCanary_failureRetentionPeriodInDays :: Lens' UpdateCanary (Maybe Natural)

The number of days to retain data about failed runs of this canary.

updateCanary_artifactConfig :: Lens' UpdateCanary (Maybe ArtifactConfigInput)

A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

updateCanary_vpcConfig :: Lens' UpdateCanary (Maybe VpcConfigInput)

If this canary is to test an endpoint in a VPC, this structure contains information about the subnet and security groups of the VPC endpoint. For more information, see Running a Canary in a VPC.

updateCanary_visualReference :: Lens' UpdateCanary (Maybe VisualReferenceInput)

Defines the screenshots to use as the baseline for comparisons during visual monitoring comparisons during future runs of this canary. If you omit this parameter, no changes are made to any baseline screenshots that the canary might be using already.

Visual monitoring is supported only on canaries running the syn-puppeteer-node-3.2 runtime or later. For more information, see Visual monitoring and Visual monitoring blueprint

updateCanary_code :: Lens' UpdateCanary (Maybe CanaryCodeInput)

A structure that includes the entry point from which the canary should start running your script. If the script is stored in an S3 bucket, the bucket name, key, and version are also included.

updateCanary_name :: Lens' UpdateCanary Text

The name of the canary that you want to update. To find the names of your canaries, use DescribeCanaries.

You cannot change the name of a canary that has already been created.

Destructuring the Response

data UpdateCanaryResponse

See: newUpdateCanaryResponse smart constructor.

Constructors

UpdateCanaryResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateCanaryResponse

Arguments

:: Int
$sel:httpStatus:UpdateCanaryResponse'

-> UpdateCanaryResponse

Create a value of UpdateCanaryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateCanaryResponse', updateCanaryResponse_httpStatus - The response's http status code.

Response Lenses

updateCanaryResponse_httpStatus :: Lens' UpdateCanaryResponse Int

The response's http status code.