Network.AWS.Textract.GetDocumentAnalysis

Description

Gets the results for an Amazon Textract asynchronous operation that analyzes text in a document.

You start asynchronous text analysis by calling StartDocumentAnalysis, which returns a job identifier (JobId). When the text analysis operation finishes, Amazon Textract publishes a completion status to the Amazon Simple Notification Service (Amazon SNS) topic that's registered in the initial call to StartDocumentAnalysis. To get the results of the text-detection operation, first check that the status value published to the Amazon SNS topic is SUCCEEDED. If so, call GetDocumentAnalysis, and pass the job identifier (JobId) from the initial call to StartDocumentAnalysis.

GetDocumentAnalysis returns an array of Block objects. The following types of information are returned:

Form data (key-value pairs). The related information is returned in two Block objects, each of type KEY_VALUE_SET: a KEY Block object and a VALUE Block object. For example, /Name: Ana Silva Carolina contains a key and value. Name: is the key. Ana Silva Carolina/ is the value.

Table and table cell data. A TABLE Block object contains information about a detected table. A CELL Block object is returned for each cell in a table.

Lines and words of text. A LINE Block object contains one or more WORD Block objects. All lines and words that are detected in the document are returned (including text that doesn't have a relationship with the value of the StartDocumentAnalysis FeatureTypes input parameter).

Selection elements such as check boxes and option buttons (radio buttons) can be detected in form data and in tables. A SELECTION_ELEMENT Block object contains information about a selection element, including the selection status.

Use the MaxResults parameter to limit the number of blocks that are returned. If there are more results than specified in MaxResults, the value of NextToken in the operation response contains a pagination token for getting the next set of results. To get the next page of results, call GetDocumentAnalysis, and populate the NextToken request parameter with the token value that's returned from the previous call to GetDocumentAnalysis.

For more information, see Document Text Analysis.

Creating a Request

data GetDocumentAnalysis

See: newGetDocumentAnalysis smart constructor.

Constructors

GetDocumentAnalysis'

Fields

nextToken :: Maybe Text

If the previous response was incomplete (because there are more blocks to retrieve), Amazon Textract returns a pagination token in the response. You can use this pagination token to retrieve the next set of blocks.

maxResults :: Maybe Natural

The maximum number of results to return per paginated call. The largest value that you can specify is 1,000. If you specify a value greater than 1,000, a maximum of 1,000 results is returned. The default value is 1,000.

jobId :: Text

A unique identifier for the text-detection job. The JobId is returned from StartDocumentAnalysis. A JobId value is only valid for 7 days.

newGetDocumentAnalysis

Arguments

:: Text
$sel:jobId:GetDocumentAnalysis'

-> GetDocumentAnalysis

Create a value of GetDocumentAnalysis with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetDocumentAnalysis', getDocumentAnalysis_nextToken - If the previous response was incomplete (because there are more blocks to retrieve), Amazon Textract returns a pagination token in the response. You can use this pagination token to retrieve the next set of blocks.

$sel:maxResults:GetDocumentAnalysis', getDocumentAnalysis_maxResults - The maximum number of results to return per paginated call. The largest value that you can specify is 1,000. If you specify a value greater than 1,000, a maximum of 1,000 results is returned. The default value is 1,000.

$sel:jobId:GetDocumentAnalysis', getDocumentAnalysis_jobId - A unique identifier for the text-detection job. The JobId is returned from StartDocumentAnalysis. A JobId value is only valid for 7 days.

Request Lenses

getDocumentAnalysis_nextToken :: Lens' GetDocumentAnalysis (Maybe Text)

If the previous response was incomplete (because there are more blocks to retrieve), Amazon Textract returns a pagination token in the response. You can use this pagination token to retrieve the next set of blocks.

getDocumentAnalysis_maxResults :: Lens' GetDocumentAnalysis (Maybe Natural)

The maximum number of results to return per paginated call. The largest value that you can specify is 1,000. If you specify a value greater than 1,000, a maximum of 1,000 results is returned. The default value is 1,000.

getDocumentAnalysis_jobId :: Lens' GetDocumentAnalysis Text

A unique identifier for the text-detection job. The JobId is returned from StartDocumentAnalysis. A JobId value is only valid for 7 days.

Destructuring the Response

data GetDocumentAnalysisResponse

See: newGetDocumentAnalysisResponse smart constructor.

Constructors

GetDocumentAnalysisResponse'

Fields

statusMessage :: Maybe Text

Returns if the detection job could not be completed. Contains explanation for what error occured.

nextToken :: Maybe Text

If the response is truncated, Amazon Textract returns this token. You can use this token in the subsequent request to retrieve the next set of text detection results.

warnings :: Maybe [Warning]

A list of warnings that occurred during the document-analysis operation.

analyzeDocumentModelVersion :: Maybe Text

jobStatus :: Maybe JobStatus

The current status of the text detection job.

blocks :: Maybe [Block]

The results of the text-analysis operation.

documentMetadata :: Maybe DocumentMetadata

Information about a document that Amazon Textract processed. DocumentMetadata is returned in every page of paginated responses from an Amazon Textract video operation.

httpStatus :: Int

The response's http status code.

newGetDocumentAnalysisResponse

Arguments

:: Int
$sel:httpStatus:GetDocumentAnalysisResponse'

-> GetDocumentAnalysisResponse

Create a value of GetDocumentAnalysisResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusMessage:GetDocumentAnalysisResponse', getDocumentAnalysisResponse_statusMessage - Returns if the detection job could not be completed. Contains explanation for what error occured.

$sel:nextToken:GetDocumentAnalysis', getDocumentAnalysisResponse_nextToken - If the response is truncated, Amazon Textract returns this token. You can use this token in the subsequent request to retrieve the next set of text detection results.

$sel:warnings:GetDocumentAnalysisResponse', getDocumentAnalysisResponse_warnings - A list of warnings that occurred during the document-analysis operation.

$sel:analyzeDocumentModelVersion:GetDocumentAnalysisResponse', getDocumentAnalysisResponse_analyzeDocumentModelVersion -

$sel:jobStatus:GetDocumentAnalysisResponse', getDocumentAnalysisResponse_jobStatus - The current status of the text detection job.

$sel:blocks:GetDocumentAnalysisResponse', getDocumentAnalysisResponse_blocks - The results of the text-analysis operation.

$sel:documentMetadata:GetDocumentAnalysisResponse', getDocumentAnalysisResponse_documentMetadata - Information about a document that Amazon Textract processed. DocumentMetadata is returned in every page of paginated responses from an Amazon Textract video operation.

$sel:httpStatus:GetDocumentAnalysisResponse', getDocumentAnalysisResponse_httpStatus - The response's http status code.

Response Lenses

getDocumentAnalysisResponse_statusMessage :: Lens' GetDocumentAnalysisResponse (Maybe Text)

Returns if the detection job could not be completed. Contains explanation for what error occured.

getDocumentAnalysisResponse_nextToken :: Lens' GetDocumentAnalysisResponse (Maybe Text)

If the response is truncated, Amazon Textract returns this token. You can use this token in the subsequent request to retrieve the next set of text detection results.

getDocumentAnalysisResponse_warnings :: Lens' GetDocumentAnalysisResponse (Maybe [Warning])

A list of warnings that occurred during the document-analysis operation.

getDocumentAnalysisResponse_analyzeDocumentModelVersion :: Lens' GetDocumentAnalysisResponse (Maybe Text)

getDocumentAnalysisResponse_jobStatus :: Lens' GetDocumentAnalysisResponse (Maybe JobStatus)

The current status of the text detection job.

getDocumentAnalysisResponse_blocks :: Lens' GetDocumentAnalysisResponse (Maybe [Block])

The results of the text-analysis operation.

getDocumentAnalysisResponse_documentMetadata :: Lens' GetDocumentAnalysisResponse (Maybe DocumentMetadata)

Information about a document that Amazon Textract processed. DocumentMetadata is returned in every page of paginated responses from an Amazon Textract video operation.

getDocumentAnalysisResponse_httpStatus :: Lens' GetDocumentAnalysisResponse Int

The response's http status code.