Network.AWS.Textract.Lens

Description

Operations

GetDocumentTextDetection

getDocumentTextDetection_nextToken :: Lens' GetDocumentTextDetection (Maybe Text)

If the previous response was incomplete (because there are more blocks to retrieve), Amazon Textract returns a pagination token in the response. You can use this pagination token to retrieve the next set of blocks.

getDocumentTextDetection_maxResults :: Lens' GetDocumentTextDetection (Maybe Natural)

The maximum number of results to return per paginated call. The largest value you can specify is 1,000. If you specify a value greater than 1,000, a maximum of 1,000 results is returned. The default value is 1,000.

getDocumentTextDetection_jobId :: Lens' GetDocumentTextDetection Text

A unique identifier for the text detection job. The JobId is returned from StartDocumentTextDetection. A JobId value is only valid for 7 days.

getDocumentTextDetectionResponse_statusMessage :: Lens' GetDocumentTextDetectionResponse (Maybe Text)

Returns if the detection job could not be completed. Contains explanation for what error occured.

getDocumentTextDetectionResponse_nextToken :: Lens' GetDocumentTextDetectionResponse (Maybe Text)

If the response is truncated, Amazon Textract returns this token. You can use this token in the subsequent request to retrieve the next set of text-detection results.

getDocumentTextDetectionResponse_warnings :: Lens' GetDocumentTextDetectionResponse (Maybe [Warning])

A list of warnings that occurred during the text-detection operation for the document.

getDocumentTextDetectionResponse_jobStatus :: Lens' GetDocumentTextDetectionResponse (Maybe JobStatus)

The current status of the text detection job.

getDocumentTextDetectionResponse_blocks :: Lens' GetDocumentTextDetectionResponse (Maybe [Block])

The results of the text-detection operation.

getDocumentTextDetectionResponse_documentMetadata :: Lens' GetDocumentTextDetectionResponse (Maybe DocumentMetadata)

Information about a document that Amazon Textract processed. DocumentMetadata is returned in every page of paginated responses from an Amazon Textract video operation.

getDocumentTextDetectionResponse_detectDocumentTextModelVersion :: Lens' GetDocumentTextDetectionResponse (Maybe Text)

getDocumentTextDetectionResponse_httpStatus :: Lens' GetDocumentTextDetectionResponse Int

The response's http status code.

GetDocumentAnalysis

getDocumentAnalysis_nextToken :: Lens' GetDocumentAnalysis (Maybe Text)

If the previous response was incomplete (because there are more blocks to retrieve), Amazon Textract returns a pagination token in the response. You can use this pagination token to retrieve the next set of blocks.

getDocumentAnalysis_maxResults :: Lens' GetDocumentAnalysis (Maybe Natural)

The maximum number of results to return per paginated call. The largest value that you can specify is 1,000. If you specify a value greater than 1,000, a maximum of 1,000 results is returned. The default value is 1,000.

getDocumentAnalysis_jobId :: Lens' GetDocumentAnalysis Text

A unique identifier for the text-detection job. The JobId is returned from StartDocumentAnalysis. A JobId value is only valid for 7 days.

getDocumentAnalysisResponse_statusMessage :: Lens' GetDocumentAnalysisResponse (Maybe Text)

Returns if the detection job could not be completed. Contains explanation for what error occured.

getDocumentAnalysisResponse_nextToken :: Lens' GetDocumentAnalysisResponse (Maybe Text)

If the response is truncated, Amazon Textract returns this token. You can use this token in the subsequent request to retrieve the next set of text detection results.

getDocumentAnalysisResponse_warnings :: Lens' GetDocumentAnalysisResponse (Maybe [Warning])

A list of warnings that occurred during the document-analysis operation.

getDocumentAnalysisResponse_analyzeDocumentModelVersion :: Lens' GetDocumentAnalysisResponse (Maybe Text)

getDocumentAnalysisResponse_jobStatus :: Lens' GetDocumentAnalysisResponse (Maybe JobStatus)

The current status of the text detection job.

getDocumentAnalysisResponse_blocks :: Lens' GetDocumentAnalysisResponse (Maybe [Block])

The results of the text-analysis operation.

getDocumentAnalysisResponse_documentMetadata :: Lens' GetDocumentAnalysisResponse (Maybe DocumentMetadata)

Information about a document that Amazon Textract processed. DocumentMetadata is returned in every page of paginated responses from an Amazon Textract video operation.

getDocumentAnalysisResponse_httpStatus :: Lens' GetDocumentAnalysisResponse Int

The response's http status code.

StartDocumentAnalysis

startDocumentAnalysis_notificationChannel :: Lens' StartDocumentAnalysis (Maybe NotificationChannel)

The Amazon SNS topic ARN that you want Amazon Textract to publish the completion status of the operation to.

startDocumentAnalysis_outputConfig :: Lens' StartDocumentAnalysis (Maybe OutputConfig)

Sets if the output will go to a customer defined bucket. By default, Amazon Textract will save the results internally to be accessed by the GetDocumentAnalysis operation.

startDocumentAnalysis_kmsKeyId :: Lens' StartDocumentAnalysis (Maybe Text)

The KMS key used to encrypt the inference results. This can be in either Key ID or Key Alias format. When a KMS key is provided, the KMS key will be used for server-side encryption of the objects in the customer bucket. When this parameter is not enabled, the result will be encrypted server side,using SSE-S3.

startDocumentAnalysis_clientRequestToken :: Lens' StartDocumentAnalysis (Maybe Text)

The idempotent token that you use to identify the start request. If you use the same token with multiple StartDocumentAnalysis requests, the same JobId is returned. Use ClientRequestToken to prevent the same job from being accidentally started more than once. For more information, see Calling Amazon Textract Asynchronous Operations.

startDocumentAnalysis_jobTag :: Lens' StartDocumentAnalysis (Maybe Text)

An identifier that you specify that's included in the completion notification published to the Amazon SNS topic. For example, you can use JobTag to identify the type of document that the completion notification corresponds to (such as a tax form or a receipt).

startDocumentAnalysis_documentLocation :: Lens' StartDocumentAnalysis DocumentLocation

The location of the document to be processed.

startDocumentAnalysis_featureTypes :: Lens' StartDocumentAnalysis [FeatureType]

A list of the types of analysis to perform. Add TABLES to the list to return information about the tables that are detected in the input document. Add FORMS to return detected form data. To perform both types of analysis, add TABLES and FORMS to FeatureTypes. All lines and words detected in the document are included in the response (including text that isn't related to the value of FeatureTypes).

startDocumentAnalysisResponse_jobId :: Lens' StartDocumentAnalysisResponse (Maybe Text)

The identifier for the document text detection job. Use JobId to identify the job in a subsequent call to GetDocumentAnalysis. A JobId value is only valid for 7 days.

startDocumentAnalysisResponse_httpStatus :: Lens' StartDocumentAnalysisResponse Int

The response's http status code.

StartDocumentTextDetection

startDocumentTextDetection_notificationChannel :: Lens' StartDocumentTextDetection (Maybe NotificationChannel)

The Amazon SNS topic ARN that you want Amazon Textract to publish the completion status of the operation to.

startDocumentTextDetection_outputConfig :: Lens' StartDocumentTextDetection (Maybe OutputConfig)

Sets if the output will go to a customer defined bucket. By default Amazon Textract will save the results internally to be accessed with the GetDocumentTextDetection operation.

startDocumentTextDetection_kmsKeyId :: Lens' StartDocumentTextDetection (Maybe Text)

The KMS key used to encrypt the inference results. This can be in either Key ID or Key Alias format. When a KMS key is provided, the KMS key will be used for server-side encryption of the objects in the customer bucket. When this parameter is not enabled, the result will be encrypted server side,using SSE-S3.

startDocumentTextDetection_clientRequestToken :: Lens' StartDocumentTextDetection (Maybe Text)

The idempotent token that's used to identify the start request. If you use the same token with multiple StartDocumentTextDetection requests, the same JobId is returned. Use ClientRequestToken to prevent the same job from being accidentally started more than once. For more information, see Calling Amazon Textract Asynchronous Operations.

startDocumentTextDetection_jobTag :: Lens' StartDocumentTextDetection (Maybe Text)

An identifier that you specify that's included in the completion notification published to the Amazon SNS topic. For example, you can use JobTag to identify the type of document that the completion notification corresponds to (such as a tax form or a receipt).

startDocumentTextDetection_documentLocation :: Lens' StartDocumentTextDetection DocumentLocation

The location of the document to be processed.

startDocumentTextDetectionResponse_jobId :: Lens' StartDocumentTextDetectionResponse (Maybe Text)

The identifier of the text detection job for the document. Use JobId to identify the job in a subsequent call to GetDocumentTextDetection. A JobId value is only valid for 7 days.

startDocumentTextDetectionResponse_httpStatus :: Lens' StartDocumentTextDetectionResponse Int

The response's http status code.

AnalyzeExpense

analyzeExpense_document :: Lens' AnalyzeExpense Document

Undocumented member.

analyzeExpenseResponse_documentMetadata :: Lens' AnalyzeExpenseResponse (Maybe DocumentMetadata)

Undocumented member.

analyzeExpenseResponse_expenseDocuments :: Lens' AnalyzeExpenseResponse (Maybe [ExpenseDocument])

The expenses detected by Amazon Textract.

analyzeExpenseResponse_httpStatus :: Lens' AnalyzeExpenseResponse Int

The response's http status code.

AnalyzeDocument

analyzeDocument_humanLoopConfig :: Lens' AnalyzeDocument (Maybe HumanLoopConfig)

Sets the configuration for the human in the loop workflow for analyzing documents.

analyzeDocument_document :: Lens' AnalyzeDocument Document

The input document as base64-encoded bytes or an Amazon S3 object. If you use the AWS CLI to call Amazon Textract operations, you can't pass image bytes. The document must be an image in JPEG or PNG format.

If you're using an AWS SDK to call Amazon Textract, you might not need to base64-encode image bytes that are passed using the Bytes field.

analyzeDocument_featureTypes :: Lens' AnalyzeDocument [FeatureType]

A list of the types of analysis to perform. Add TABLES to the list to return information about the tables that are detected in the input document. Add FORMS to return detected form data. To perform both types of analysis, add TABLES and FORMS to FeatureTypes. All lines and words detected in the document are included in the response (including text that isn't related to the value of FeatureTypes).

analyzeDocumentResponse_analyzeDocumentModelVersion :: Lens' AnalyzeDocumentResponse (Maybe Text)

The version of the model used to analyze the document.

analyzeDocumentResponse_blocks :: Lens' AnalyzeDocumentResponse (Maybe [Block])

The items that are detected and analyzed by AnalyzeDocument.

analyzeDocumentResponse_documentMetadata :: Lens' AnalyzeDocumentResponse (Maybe DocumentMetadata)

Metadata about the analyzed document. An example is the number of pages.

analyzeDocumentResponse_humanLoopActivationOutput :: Lens' AnalyzeDocumentResponse (Maybe HumanLoopActivationOutput)

Shows the results of the human in the loop evaluation.

analyzeDocumentResponse_httpStatus :: Lens' AnalyzeDocumentResponse Int

The response's http status code.

DetectDocumentText

detectDocumentText_document :: Lens' DetectDocumentText Document

The input document as base64-encoded bytes or an Amazon S3 object. If you use the AWS CLI to call Amazon Textract operations, you can't pass image bytes. The document must be an image in JPEG or PNG format.

If you're using an AWS SDK to call Amazon Textract, you might not need to base64-encode image bytes that are passed using the Bytes field.

detectDocumentTextResponse_blocks :: Lens' DetectDocumentTextResponse (Maybe [Block])

An array of Block objects that contain the text that's detected in the document.

detectDocumentTextResponse_documentMetadata :: Lens' DetectDocumentTextResponse (Maybe DocumentMetadata)

Metadata about the document. It contains the number of pages that are detected in the document.

detectDocumentTextResponse_detectDocumentTextModelVersion :: Lens' DetectDocumentTextResponse (Maybe Text)

detectDocumentTextResponse_httpStatus :: Lens' DetectDocumentTextResponse Int

The response's http status code.

Types

Block

block_relationships :: Lens' Block (Maybe [Relationship])

A list of child blocks of the current block. For example, a LINE object has child blocks for each WORD block that's part of the line of text. There aren't Relationship objects in the list for relationships that don't exist, such as when the current block has no child blocks. The list size can be the following:

0 - The block has no child blocks.

1 - The block has child blocks.

block_selectionStatus :: Lens' Block (Maybe SelectionStatus)

The selection status of a selection element, such as an option button or check box.

block_blockType :: Lens' Block (Maybe BlockType)

The type of text item that's recognized. In operations for text detection, the following types are returned:

PAGE - Contains a list of the LINE Block objects that are detected on a document page.

WORD - A word detected on a document page. A word is one or more ISO basic Latin script characters that aren't separated by spaces.

LINE - A string of tab-delimited, contiguous words that are detected on a document page.

In text analysis operations, the following types are returned:

PAGE - Contains a list of child Block objects that are detected on a document page.

KEY_VALUE_SET - Stores the KEY and VALUE Block objects for linked text that's detected on a document page. Use the EntityType field to determine if a KEY_VALUE_SET object is a KEY Block object or a VALUE Block object.

WORD - A word that's detected on a document page. A word is one or more ISO basic Latin script characters that aren't separated by spaces.

LINE - A string of tab-delimited, contiguous words that are detected on a document page.

TABLE - A table that's detected on a document page. A table is grid-based information with two or more rows or columns, with a cell span of one row and one column each.

CELL - A cell within a detected table. The cell is the parent of the block that contains the text in the cell.

SELECTION_ELEMENT - A selection element such as an option button (radio button) or a check box that's detected on a document page. Use the value of SelectionStatus to determine the status of the selection element.

block_rowSpan :: Lens' Block (Maybe Natural)

The number of rows that a table cell spans. Currently this value is always 1, even if the number of rows spanned is greater than 1. RowSpan isn't returned by DetectDocumentText and GetDocumentTextDetection.

block_page :: Lens' Block (Maybe Natural)

The page on which a block was detected. Page is returned by asynchronous operations. Page values greater than 1 are only returned for multipage documents that are in PDF format. A scanned image (JPEG/PNG), even if it contains multiple document pages, is considered to be a single-page document. The value of Page is always 1. Synchronous operations don't return Page because every input document is considered to be a single-page document.

block_id :: Lens' Block (Maybe Text)

The identifier for the recognized text. The identifier is only unique for a single operation.

block_textType :: Lens' Block (Maybe TextType)

The kind of text that Amazon Textract has detected. Can check for handwritten text and printed text.

block_columnSpan :: Lens' Block (Maybe Natural)

The number of columns that a table cell spans. Currently this value is always 1, even if the number of columns spanned is greater than 1. ColumnSpan isn't returned by DetectDocumentText and GetDocumentTextDetection.

block_rowIndex :: Lens' Block (Maybe Natural)

The row in which a table cell is located. The first row position is 1. RowIndex isn't returned by DetectDocumentText and GetDocumentTextDetection.

block_confidence :: Lens' Block (Maybe Double)

The confidence score that Amazon Textract has in the accuracy of the recognized text and the accuracy of the geometry points around the recognized text.

block_columnIndex :: Lens' Block (Maybe Natural)

The column in which a table cell appears. The first column position is 1. ColumnIndex isn't returned by DetectDocumentText and GetDocumentTextDetection.

block_entityTypes :: Lens' Block (Maybe [EntityType])

The type of entity. The following can be returned:

KEY - An identifier for a field on the document.

VALUE - The field text.

EntityTypes isn't returned by DetectDocumentText and GetDocumentTextDetection.

block_text :: Lens' Block (Maybe Text)

The word or line of text that's recognized by Amazon Textract.

block_geometry :: Lens' Block (Maybe Geometry)

The location of the recognized text on the image. It includes an axis-aligned, coarse bounding box that surrounds the text, and a finer-grain polygon for more accurate spatial information.

BoundingBox

boundingBox_height :: Lens' BoundingBox (Maybe Double)

The height of the bounding box as a ratio of the overall document page height.

boundingBox_width :: Lens' BoundingBox (Maybe Double)

The width of the bounding box as a ratio of the overall document page width.

boundingBox_left :: Lens' BoundingBox (Maybe Double)

The left coordinate of the bounding box as a ratio of overall document page width.

boundingBox_top :: Lens' BoundingBox (Maybe Double)

The top coordinate of the bounding box as a ratio of overall document page height.

Document

document_bytes :: Lens' Document (Maybe ByteString)

A blob of base64-encoded document bytes. The maximum size of a document that's provided in a blob of bytes is 5 MB. The document bytes must be in PNG or JPEG format.

If you're using an AWS SDK to call Amazon Textract, you might not need to base64-encode image bytes passed using the Bytes field.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

document_s3Object :: Lens' Document (Maybe S3Object)

Identifies an S3 object as the document source. The maximum size of a document that's stored in an S3 bucket is 5 MB.

DocumentLocation

documentLocation_s3Object :: Lens' DocumentLocation (Maybe S3Object)

The Amazon S3 bucket that contains the input document.

DocumentMetadata

documentMetadata_pages :: Lens' DocumentMetadata (Maybe Natural)

The number of pages that are detected in the document.

ExpenseDetection

expenseDetection_confidence :: Lens' ExpenseDetection (Maybe Double)

The confidence in detection, as a percentage

expenseDetection_text :: Lens' ExpenseDetection (Maybe Text)

The word or line of text recognized by Amazon Textract

expenseDetection_geometry :: Lens' ExpenseDetection (Maybe Geometry)

Undocumented member.

ExpenseDocument

expenseDocument_lineItemGroups :: Lens' ExpenseDocument (Maybe [LineItemGroup])

Information detected on each table of a document, seperated into LineItems.

expenseDocument_expenseIndex :: Lens' ExpenseDocument (Maybe Natural)

Denotes which invoice or receipt in the document the information is coming from. First document will be 1, the second 2, and so on.

expenseDocument_summaryFields :: Lens' ExpenseDocument (Maybe [ExpenseField])

Any information found outside of a table by Amazon Textract.

ExpenseField

expenseField_labelDetection :: Lens' ExpenseField (Maybe ExpenseDetection)

The explicitly stated label of a detected element.

expenseField_valueDetection :: Lens' ExpenseField (Maybe ExpenseDetection)

The value of a detected element. Present in explicit and implicit elements.

expenseField_pageNumber :: Lens' ExpenseField (Maybe Natural)

The page number the value was detected on.

expenseField_type :: Lens' ExpenseField (Maybe ExpenseType)

The implied label of a detected element. Present alongside LabelDetection for explicit elements.

ExpenseType

expenseType_confidence :: Lens' ExpenseType (Maybe Double)

The confidence of accuracy, as a percentage.

expenseType_text :: Lens' ExpenseType (Maybe Text)

The word or line of text detected by Amazon Textract.

Geometry

geometry_polygon :: Lens' Geometry (Maybe [Point])

Within the bounding box, a fine-grained polygon around the recognized item.

geometry_boundingBox :: Lens' Geometry (Maybe BoundingBox)

An axis-aligned coarse representation of the location of the recognized item on the document page.

HumanLoopActivationOutput

humanLoopActivationOutput_humanLoopActivationReasons :: Lens' HumanLoopActivationOutput (Maybe (NonEmpty Text))

Shows if and why human review was needed.

humanLoopActivationOutput_humanLoopArn :: Lens' HumanLoopActivationOutput (Maybe Text)

The Amazon Resource Name (ARN) of the HumanLoop created.

humanLoopActivationOutput_humanLoopActivationConditionsEvaluationResults :: Lens' HumanLoopActivationOutput (Maybe Text)

Shows the result of condition evaluations, including those conditions which activated a human review.

HumanLoopConfig

humanLoopConfig_dataAttributes :: Lens' HumanLoopConfig (Maybe HumanLoopDataAttributes)

Sets attributes of the input data.

humanLoopConfig_humanLoopName :: Lens' HumanLoopConfig Text

The name of the human workflow used for this image. This should be kept unique within a region.

humanLoopConfig_flowDefinitionArn :: Lens' HumanLoopConfig Text

The Amazon Resource Name (ARN) of the flow definition.

HumanLoopDataAttributes

humanLoopDataAttributes_contentClassifiers :: Lens' HumanLoopDataAttributes (Maybe [ContentClassifier])

Sets whether the input image is free of personally identifiable information or adult content.

LineItemFields

lineItemFields_lineItemExpenseFields :: Lens' LineItemFields (Maybe [ExpenseField])

ExpenseFields used to show information from detected lines on a table.

LineItemGroup

lineItemGroup_lineItems :: Lens' LineItemGroup (Maybe [LineItemFields])

The breakdown of information on a particular line of a table.

lineItemGroup_lineItemGroupIndex :: Lens' LineItemGroup (Maybe Natural)

The number used to identify a specific table in a document. The first table encountered will have a LineItemGroupIndex of 1, the second 2, etc.

NotificationChannel

notificationChannel_sNSTopicArn :: Lens' NotificationChannel Text

The Amazon SNS topic that Amazon Textract posts the completion status to.

notificationChannel_roleArn :: Lens' NotificationChannel Text

The Amazon Resource Name (ARN) of an IAM role that gives Amazon Textract publishing permissions to the Amazon SNS topic.

OutputConfig

outputConfig_s3Prefix :: Lens' OutputConfig (Maybe Text)

The prefix of the object key that the output will be saved to. When not enabled, the prefix will be “textract_output".

outputConfig_s3Bucket :: Lens' OutputConfig Text

The name of the bucket your output will go to.

Point

point_y :: Lens' Point (Maybe Double)

The value of the Y coordinate for a point on a Polygon.

point_x :: Lens' Point (Maybe Double)

The value of the X coordinate for a point on a Polygon.

Relationship

relationship_ids :: Lens' Relationship (Maybe [Text])

An array of IDs for related blocks. You can get the type of the relationship from the Type element.

relationship_type :: Lens' Relationship (Maybe RelationshipType)

The type of relationship that the blocks in the IDs array have with the current block. The relationship can be VALUE or CHILD. A relationship of type VALUE is a list that contains the ID of the VALUE block that's associated with the KEY of a key-value pair. A relationship of type CHILD is a list of IDs that identify WORD blocks in the case of lines Cell blocks in the case of Tables, and WORD blocks in the case of Selection Elements.

S3Object

s3Object_version :: Lens' S3Object (Maybe Text)

If the bucket has versioning enabled, you can specify the object version.

s3Object_name :: Lens' S3Object (Maybe Text)

The file name of the input document. Synchronous operations can use image files that are in JPEG or PNG format. Asynchronous operations also support PDF format files.

s3Object_bucket :: Lens' S3Object (Maybe Text)

The name of the S3 bucket. Note that the # character is not valid in the file name.

Warning

warning_pages :: Lens' Warning (Maybe [Natural])

A list of the pages that the warning applies to.

warning_errorCode :: Lens' Warning (Maybe Text)

The error code for the warning.