Network.AWS.Textract.Types

Description

Service Configuration

defaultService :: Service

API version 2018-06-27 of the Amazon Textract SDK configuration.

Errors

_BadDocumentException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Textract isn't able to read the document. For more information on the document limits in Amazon Textract, see limits.

_DocumentTooLargeException :: AsError a => Getting (First ServiceError) a ServiceError

The document can't be processed because it's too large. The maximum document size for synchronous operations 10 MB. The maximum document size for asynchronous operations is 500 MB for PDF files.

_HumanLoopQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates you have exceeded the maximum number of active human in the loop workflows available

_InvalidS3ObjectException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Textract is unable to access the S3 object that's specified in the request. for more information, Configure Access to Amazon S3 For troubleshooting information, see Troubleshooting Amazon S3

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Textract experienced a service issue. Try your call again.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Textract is temporarily unable to process the request. Try your call again.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

An input parameter violated a constraint. For example, in synchronous operations, an InvalidParameterException exception occurs when neither of the S3Object or Bytes values are supplied in the Document request parameter. Validate your parameter before calling the API operation again.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You aren't authorized to perform the action. Use the Amazon Resource Name (ARN) of an authorized user or IAM role to perform the operation.

_InvalidKMSKeyException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates you do not have decrypt permissions with the KMS key entered, or the KMS key was entered incorrectly.

_InvalidJobIdException :: AsError a => Getting (First ServiceError) a ServiceError

An invalid job identifier was passed to GetDocumentAnalysis or to GetDocumentAnalysis.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

An Amazon Textract service limit was exceeded. For example, if you start too many asynchronous jobs concurrently, calls to start operations (StartDocumentTextDetection, for example) raise a LimitExceededException exception (HTTP status code: 400) until the number of concurrently running jobs is below the Amazon Textract service limit.

_ProvisionedThroughputExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The number of requests exceeded your throughput limit. If you want to increase this limit, contact Amazon Textract.

_IdempotentParameterMismatchException :: AsError a => Getting (First ServiceError) a ServiceError

A ClientRequestToken input parameter was reused with an operation, but at least one of the other input parameters is different from the previous call to the operation.

_UnsupportedDocumentException :: AsError a => Getting (First ServiceError) a ServiceError

The format of the input document isn't supported. Documents for synchronous operations can be in PNG or JPEG format. Documents for asynchronous operations can also be in PDF format.

BlockType

newtype BlockType

Constructors

BlockType'

Fields

fromBlockType :: Text

Bundled Patterns

pattern BlockType_CELL :: BlockType

pattern BlockType_KEY_VALUE_SET :: BlockType

pattern BlockType_LINE :: BlockType

pattern BlockType_PAGE :: BlockType

pattern BlockType_SELECTION_ELEMENT :: BlockType

pattern BlockType_TABLE :: BlockType

pattern BlockType_WORD :: BlockType

ContentClassifier

newtype ContentClassifier

Constructors

ContentClassifier'

Fields

fromContentClassifier :: Text

Bundled Patterns

pattern ContentClassifier_FreeOfAdultContent :: ContentClassifier

pattern ContentClassifier_FreeOfPersonallyIdentifiableInformation :: ContentClassifier

EntityType

newtype EntityType

Constructors

EntityType'

Fields

fromEntityType :: Text

Bundled Patterns

pattern EntityType_KEY :: EntityType

pattern EntityType_VALUE :: EntityType

FeatureType

newtype FeatureType

Constructors

FeatureType'

Fields

fromFeatureType :: Text

Bundled Patterns

pattern FeatureType_FORMS :: FeatureType

pattern FeatureType_TABLES :: FeatureType

JobStatus

newtype JobStatus

Constructors

JobStatus'

Fields

fromJobStatus :: Text

Bundled Patterns

pattern JobStatus_FAILED :: JobStatus

pattern JobStatus_IN_PROGRESS :: JobStatus

pattern JobStatus_PARTIAL_SUCCESS :: JobStatus

pattern JobStatus_SUCCEEDED :: JobStatus

RelationshipType

newtype RelationshipType

Constructors

RelationshipType'

Fields

fromRelationshipType :: Text

Bundled Patterns

pattern RelationshipType_CHILD :: RelationshipType

pattern RelationshipType_COMPLEX_FEATURES :: RelationshipType

pattern RelationshipType_VALUE :: RelationshipType

SelectionStatus

newtype SelectionStatus

Constructors

SelectionStatus'

Fields

fromSelectionStatus :: Text

Bundled Patterns

pattern SelectionStatus_NOT_SELECTED :: SelectionStatus

pattern SelectionStatus_SELECTED :: SelectionStatus

TextType

newtype TextType

Constructors

TextType'

Fields

fromTextType :: Text

Bundled Patterns

pattern TextType_HANDWRITING :: TextType

pattern TextType_PRINTED :: TextType

Block

data Block

A Block represents items that are recognized in a document within a group of pixels close to each other. The information returned in a Block object depends on the type of operation. In text detection for documents (for example DetectDocumentText), you get information about the detected words and lines of text. In text analysis (for example AnalyzeDocument), you can also get information about the fields, tables, and selection elements that are detected in the document.

An array of Block objects is returned by both synchronous and asynchronous operations. In synchronous operations, such as DetectDocumentText, the array of Block objects is the entire set of results. In asynchronous operations, such as GetDocumentAnalysis, the array is returned over one or more responses.

For more information, see How Amazon Textract Works.

See: newBlock smart constructor.

Constructors

Block'

Fields

relationships :: Maybe [Relationship]

A list of child blocks of the current block. For example, a LINE object has child blocks for each WORD block that's part of the line of text. There aren't Relationship objects in the list for relationships that don't exist, such as when the current block has no child blocks. The list size can be the following:

0 - The block has no child blocks.

1 - The block has child blocks.

selectionStatus :: Maybe SelectionStatus

The selection status of a selection element, such as an option button or check box.

blockType :: Maybe BlockType

The type of text item that's recognized. In operations for text detection, the following types are returned:

PAGE - Contains a list of the LINE Block objects that are detected on a document page.

WORD - A word detected on a document page. A word is one or more ISO basic Latin script characters that aren't separated by spaces.

LINE - A string of tab-delimited, contiguous words that are detected on a document page.

In text analysis operations, the following types are returned:

PAGE - Contains a list of child Block objects that are detected on a document page.

KEY_VALUE_SET - Stores the KEY and VALUE Block objects for linked text that's detected on a document page. Use the EntityType field to determine if a KEY_VALUE_SET object is a KEY Block object or a VALUE Block object.

WORD - A word that's detected on a document page. A word is one or more ISO basic Latin script characters that aren't separated by spaces.

LINE - A string of tab-delimited, contiguous words that are detected on a document page.

TABLE - A table that's detected on a document page. A table is grid-based information with two or more rows or columns, with a cell span of one row and one column each.

CELL - A cell within a detected table. The cell is the parent of the block that contains the text in the cell.

SELECTION_ELEMENT - A selection element such as an option button (radio button) or a check box that's detected on a document page. Use the value of SelectionStatus to determine the status of the selection element.

rowSpan :: Maybe Natural

The number of rows that a table cell spans. Currently this value is always 1, even if the number of rows spanned is greater than 1. RowSpan isn't returned by DetectDocumentText and GetDocumentTextDetection.

page :: Maybe Natural

The page on which a block was detected. Page is returned by asynchronous operations. Page values greater than 1 are only returned for multipage documents that are in PDF format. A scanned image (JPEG/PNG), even if it contains multiple document pages, is considered to be a single-page document. The value of Page is always 1. Synchronous operations don't return Page because every input document is considered to be a single-page document.

id :: Maybe Text

The identifier for the recognized text. The identifier is only unique for a single operation.

textType :: Maybe TextType

The kind of text that Amazon Textract has detected. Can check for handwritten text and printed text.

columnSpan :: Maybe Natural

The number of columns that a table cell spans. Currently this value is always 1, even if the number of columns spanned is greater than 1. ColumnSpan isn't returned by DetectDocumentText and GetDocumentTextDetection.

rowIndex :: Maybe Natural

The row in which a table cell is located. The first row position is 1. RowIndex isn't returned by DetectDocumentText and GetDocumentTextDetection.

confidence :: Maybe Double

The confidence score that Amazon Textract has in the accuracy of the recognized text and the accuracy of the geometry points around the recognized text.

columnIndex :: Maybe Natural

The column in which a table cell appears. The first column position is 1. ColumnIndex isn't returned by DetectDocumentText and GetDocumentTextDetection.

entityTypes :: Maybe [EntityType]

The type of entity. The following can be returned:

KEY - An identifier for a field on the document.

VALUE - The field text.

EntityTypes isn't returned by DetectDocumentText and GetDocumentTextDetection.

text :: Maybe Text

The word or line of text that's recognized by Amazon Textract.

geometry :: Maybe Geometry

The location of the recognized text on the image. It includes an axis-aligned, coarse bounding box that surrounds the text, and a finer-grain polygon for more accurate spatial information.

newBlock :: Block

Create a value of Block with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:relationships:Block', block_relationships - A list of child blocks of the current block. For example, a LINE object has child blocks for each WORD block that's part of the line of text. There aren't Relationship objects in the list for relationships that don't exist, such as when the current block has no child blocks. The list size can be the following:

0 - The block has no child blocks.

1 - The block has child blocks.

$sel:selectionStatus:Block', block_selectionStatus - The selection status of a selection element, such as an option button or check box.

$sel:blockType:Block', block_blockType - The type of text item that's recognized. In operations for text detection, the following types are returned:

PAGE - Contains a list of the LINE Block objects that are detected on a document page.

WORD - A word detected on a document page. A word is one or more ISO basic Latin script characters that aren't separated by spaces.

LINE - A string of tab-delimited, contiguous words that are detected on a document page.

In text analysis operations, the following types are returned:

PAGE - Contains a list of child Block objects that are detected on a document page.

KEY_VALUE_SET - Stores the KEY and VALUE Block objects for linked text that's detected on a document page. Use the EntityType field to determine if a KEY_VALUE_SET object is a KEY Block object or a VALUE Block object.

WORD - A word that's detected on a document page. A word is one or more ISO basic Latin script characters that aren't separated by spaces.

LINE - A string of tab-delimited, contiguous words that are detected on a document page.

TABLE - A table that's detected on a document page. A table is grid-based information with two or more rows or columns, with a cell span of one row and one column each.

CELL - A cell within a detected table. The cell is the parent of the block that contains the text in the cell.

SELECTION_ELEMENT - A selection element such as an option button (radio button) or a check box that's detected on a document page. Use the value of SelectionStatus to determine the status of the selection element.

$sel:rowSpan:Block', block_rowSpan - The number of rows that a table cell spans. Currently this value is always 1, even if the number of rows spanned is greater than 1. RowSpan isn't returned by DetectDocumentText and GetDocumentTextDetection.

$sel:page:Block', block_page - The page on which a block was detected. Page is returned by asynchronous operations. Page values greater than 1 are only returned for multipage documents that are in PDF format. A scanned image (JPEG/PNG), even if it contains multiple document pages, is considered to be a single-page document. The value of Page is always 1. Synchronous operations don't return Page because every input document is considered to be a single-page document.

$sel:id:Block', block_id - The identifier for the recognized text. The identifier is only unique for a single operation.

$sel:textType:Block', block_textType - The kind of text that Amazon Textract has detected. Can check for handwritten text and printed text.

$sel:columnSpan:Block', block_columnSpan - The number of columns that a table cell spans. Currently this value is always 1, even if the number of columns spanned is greater than 1. ColumnSpan isn't returned by DetectDocumentText and GetDocumentTextDetection.

$sel:rowIndex:Block', block_rowIndex - The row in which a table cell is located. The first row position is 1. RowIndex isn't returned by DetectDocumentText and GetDocumentTextDetection.

$sel:confidence:Block', block_confidence - The confidence score that Amazon Textract has in the accuracy of the recognized text and the accuracy of the geometry points around the recognized text.

$sel:columnIndex:Block', block_columnIndex - The column in which a table cell appears. The first column position is 1. ColumnIndex isn't returned by DetectDocumentText and GetDocumentTextDetection.

$sel:entityTypes:Block', block_entityTypes - The type of entity. The following can be returned:

KEY - An identifier for a field on the document.

VALUE - The field text.

EntityTypes isn't returned by DetectDocumentText and GetDocumentTextDetection.

$sel:text:Block', block_text - The word or line of text that's recognized by Amazon Textract.

$sel:geometry:Block', block_geometry - The location of the recognized text on the image. It includes an axis-aligned, coarse bounding box that surrounds the text, and a finer-grain polygon for more accurate spatial information.

block_relationships :: Lens' Block (Maybe [Relationship])

A list of child blocks of the current block. For example, a LINE object has child blocks for each WORD block that's part of the line of text. There aren't Relationship objects in the list for relationships that don't exist, such as when the current block has no child blocks. The list size can be the following:

0 - The block has no child blocks.

1 - The block has child blocks.

block_selectionStatus :: Lens' Block (Maybe SelectionStatus)

The selection status of a selection element, such as an option button or check box.

block_blockType :: Lens' Block (Maybe BlockType)

The type of text item that's recognized. In operations for text detection, the following types are returned:

PAGE - Contains a list of the LINE Block objects that are detected on a document page.

WORD - A word detected on a document page. A word is one or more ISO basic Latin script characters that aren't separated by spaces.

LINE - A string of tab-delimited, contiguous words that are detected on a document page.

In text analysis operations, the following types are returned:

PAGE - Contains a list of child Block objects that are detected on a document page.

KEY_VALUE_SET - Stores the KEY and VALUE Block objects for linked text that's detected on a document page. Use the EntityType field to determine if a KEY_VALUE_SET object is a KEY Block object or a VALUE Block object.

WORD - A word that's detected on a document page. A word is one or more ISO basic Latin script characters that aren't separated by spaces.

LINE - A string of tab-delimited, contiguous words that are detected on a document page.

TABLE - A table that's detected on a document page. A table is grid-based information with two or more rows or columns, with a cell span of one row and one column each.

CELL - A cell within a detected table. The cell is the parent of the block that contains the text in the cell.

SELECTION_ELEMENT - A selection element such as an option button (radio button) or a check box that's detected on a document page. Use the value of SelectionStatus to determine the status of the selection element.

block_rowSpan :: Lens' Block (Maybe Natural)

The number of rows that a table cell spans. Currently this value is always 1, even if the number of rows spanned is greater than 1. RowSpan isn't returned by DetectDocumentText and GetDocumentTextDetection.

block_page :: Lens' Block (Maybe Natural)

The page on which a block was detected. Page is returned by asynchronous operations. Page values greater than 1 are only returned for multipage documents that are in PDF format. A scanned image (JPEG/PNG), even if it contains multiple document pages, is considered to be a single-page document. The value of Page is always 1. Synchronous operations don't return Page because every input document is considered to be a single-page document.

block_id :: Lens' Block (Maybe Text)

The identifier for the recognized text. The identifier is only unique for a single operation.

block_textType :: Lens' Block (Maybe TextType)

The kind of text that Amazon Textract has detected. Can check for handwritten text and printed text.

block_columnSpan :: Lens' Block (Maybe Natural)

The number of columns that a table cell spans. Currently this value is always 1, even if the number of columns spanned is greater than 1. ColumnSpan isn't returned by DetectDocumentText and GetDocumentTextDetection.

block_rowIndex :: Lens' Block (Maybe Natural)

The row in which a table cell is located. The first row position is 1. RowIndex isn't returned by DetectDocumentText and GetDocumentTextDetection.

block_confidence :: Lens' Block (Maybe Double)

The confidence score that Amazon Textract has in the accuracy of the recognized text and the accuracy of the geometry points around the recognized text.

block_columnIndex :: Lens' Block (Maybe Natural)

The column in which a table cell appears. The first column position is 1. ColumnIndex isn't returned by DetectDocumentText and GetDocumentTextDetection.

block_entityTypes :: Lens' Block (Maybe [EntityType])

The type of entity. The following can be returned:

KEY - An identifier for a field on the document.

VALUE - The field text.

EntityTypes isn't returned by DetectDocumentText and GetDocumentTextDetection.

block_text :: Lens' Block (Maybe Text)

The word or line of text that's recognized by Amazon Textract.

block_geometry :: Lens' Block (Maybe Geometry)

The location of the recognized text on the image. It includes an axis-aligned, coarse bounding box that surrounds the text, and a finer-grain polygon for more accurate spatial information.

BoundingBox

data BoundingBox

The bounding box around the detected page, text, key-value pair, table, table cell, or selection element on a document page. The left (x-coordinate) and top (y-coordinate) are coordinates that represent the top and left sides of the bounding box. Note that the upper-left corner of the image is the origin (0,0).

The top and left values returned are ratios of the overall document page size. For example, if the input image is 700 x 200 pixels, and the top-left coordinate of the bounding box is 350 x 50 pixels, the API returns a left value of 0.5 (350/700) and a top value of 0.25 (50/200).

The width and height values represent the dimensions of the bounding box as a ratio of the overall document page dimension. For example, if the document page size is 700 x 200 pixels, and the bounding box width is 70 pixels, the width returned is 0.1.

See: newBoundingBox smart constructor.

Constructors

BoundingBox'

Fields

height :: Maybe Double

The height of the bounding box as a ratio of the overall document page height.

width :: Maybe Double

The width of the bounding box as a ratio of the overall document page width.

left :: Maybe Double

The left coordinate of the bounding box as a ratio of overall document page width.

top :: Maybe Double

The top coordinate of the bounding box as a ratio of overall document page height.

newBoundingBox :: BoundingBox

Create a value of BoundingBox with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:height:BoundingBox', boundingBox_height - The height of the bounding box as a ratio of the overall document page height.

$sel:width:BoundingBox', boundingBox_width - The width of the bounding box as a ratio of the overall document page width.

$sel:left:BoundingBox', boundingBox_left - The left coordinate of the bounding box as a ratio of overall document page width.

$sel:top:BoundingBox', boundingBox_top - The top coordinate of the bounding box as a ratio of overall document page height.

boundingBox_height :: Lens' BoundingBox (Maybe Double)

The height of the bounding box as a ratio of the overall document page height.

boundingBox_width :: Lens' BoundingBox (Maybe Double)

The width of the bounding box as a ratio of the overall document page width.

boundingBox_left :: Lens' BoundingBox (Maybe Double)

The left coordinate of the bounding box as a ratio of overall document page width.

boundingBox_top :: Lens' BoundingBox (Maybe Double)

The top coordinate of the bounding box as a ratio of overall document page height.

Document

data Document

The input document, either as bytes or as an S3 object.

You pass image bytes to an Amazon Textract API operation by using the Bytes property. For example, you would use the Bytes property to pass a document loaded from a local file system. Image bytes passed by using the Bytes property must be base64 encoded. Your code might not need to encode document file bytes if you're using an AWS SDK to call Amazon Textract API operations.

You pass images stored in an S3 bucket to an Amazon Textract API operation by using the S3Object property. Documents stored in an S3 bucket don't need to be base64 encoded.

The AWS Region for the S3 bucket that contains the S3 object must match the AWS Region that you use for Amazon Textract operations.

If you use the AWS CLI to call Amazon Textract operations, passing image bytes using the Bytes property isn't supported. You must first upload the document to an Amazon S3 bucket, and then call the operation using the S3Object property.

For Amazon Textract to process an S3 object, the user must have permission to access the S3 object.

See: newDocument smart constructor.

Constructors

Document'

Fields

bytes :: Maybe Base64

A blob of base64-encoded document bytes. The maximum size of a document that's provided in a blob of bytes is 5 MB. The document bytes must be in PNG or JPEG format.

If you're using an AWS SDK to call Amazon Textract, you might not need to base64-encode image bytes passed using the Bytes field.

s3Object :: Maybe S3Object

Identifies an S3 object as the document source. The maximum size of a document that's stored in an S3 bucket is 5 MB.

newDocument :: Document

Create a value of Document with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bytes:Document', document_bytes - A blob of base64-encoded document bytes. The maximum size of a document that's provided in a blob of bytes is 5 MB. The document bytes must be in PNG or JPEG format.

If you're using an AWS SDK to call Amazon Textract, you might not need to base64-encode image bytes passed using the Bytes field.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:s3Object:Document', document_s3Object - Identifies an S3 object as the document source. The maximum size of a document that's stored in an S3 bucket is 5 MB.

document_bytes :: Lens' Document (Maybe ByteString)

A blob of base64-encoded document bytes. The maximum size of a document that's provided in a blob of bytes is 5 MB. The document bytes must be in PNG or JPEG format.

If you're using an AWS SDK to call Amazon Textract, you might not need to base64-encode image bytes passed using the Bytes field.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

document_s3Object :: Lens' Document (Maybe S3Object)

Identifies an S3 object as the document source. The maximum size of a document that's stored in an S3 bucket is 5 MB.

DocumentLocation

data DocumentLocation

The Amazon S3 bucket that contains the document to be processed. It's used by asynchronous operations such as StartDocumentTextDetection.

The input document can be an image file in JPEG or PNG format. It can also be a file in PDF format.

See: newDocumentLocation smart constructor.

Constructors

DocumentLocation'

Fields

s3Object :: Maybe S3Object

The Amazon S3 bucket that contains the input document.

newDocumentLocation :: DocumentLocation

Create a value of DocumentLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Object:DocumentLocation', documentLocation_s3Object - The Amazon S3 bucket that contains the input document.

documentLocation_s3Object :: Lens' DocumentLocation (Maybe S3Object)

The Amazon S3 bucket that contains the input document.

DocumentMetadata

data DocumentMetadata

Information about the input document.

See: newDocumentMetadata smart constructor.

Constructors

DocumentMetadata'

Fields

pages :: Maybe Natural

The number of pages that are detected in the document.

newDocumentMetadata :: DocumentMetadata

Create a value of DocumentMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pages:DocumentMetadata', documentMetadata_pages - The number of pages that are detected in the document.

documentMetadata_pages :: Lens' DocumentMetadata (Maybe Natural)

The number of pages that are detected in the document.

ExpenseDetection

data ExpenseDetection

An object used to store information about the Value or Label detected by Amazon Textract.

See: newExpenseDetection smart constructor.

Constructors

ExpenseDetection'

Fields

confidence :: Maybe Double

The confidence in detection, as a percentage

text :: Maybe Text

The word or line of text recognized by Amazon Textract

geometry :: Maybe Geometry

newExpenseDetection :: ExpenseDetection

Create a value of ExpenseDetection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:confidence:ExpenseDetection', expenseDetection_confidence - The confidence in detection, as a percentage

$sel:text:ExpenseDetection', expenseDetection_text - The word or line of text recognized by Amazon Textract

$sel:geometry:ExpenseDetection', expenseDetection_geometry - Undocumented member.

expenseDetection_confidence :: Lens' ExpenseDetection (Maybe Double)

The confidence in detection, as a percentage

expenseDetection_text :: Lens' ExpenseDetection (Maybe Text)

The word or line of text recognized by Amazon Textract

expenseDetection_geometry :: Lens' ExpenseDetection (Maybe Geometry)

Undocumented member.

ExpenseDocument

data ExpenseDocument

The structure holding all the information returned by AnalyzeExpense

See: newExpenseDocument smart constructor.

Constructors

ExpenseDocument'

Fields

lineItemGroups :: Maybe [LineItemGroup]

Information detected on each table of a document, seperated into LineItems.

expenseIndex :: Maybe Natural

Denotes which invoice or receipt in the document the information is coming from. First document will be 1, the second 2, and so on.

summaryFields :: Maybe [ExpenseField]

Any information found outside of a table by Amazon Textract.

newExpenseDocument :: ExpenseDocument

Create a value of ExpenseDocument with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lineItemGroups:ExpenseDocument', expenseDocument_lineItemGroups - Information detected on each table of a document, seperated into LineItems.

$sel:expenseIndex:ExpenseDocument', expenseDocument_expenseIndex - Denotes which invoice or receipt in the document the information is coming from. First document will be 1, the second 2, and so on.

$sel:summaryFields:ExpenseDocument', expenseDocument_summaryFields - Any information found outside of a table by Amazon Textract.

expenseDocument_lineItemGroups :: Lens' ExpenseDocument (Maybe [LineItemGroup])

Information detected on each table of a document, seperated into LineItems.

expenseDocument_expenseIndex :: Lens' ExpenseDocument (Maybe Natural)

Denotes which invoice or receipt in the document the information is coming from. First document will be 1, the second 2, and so on.

expenseDocument_summaryFields :: Lens' ExpenseDocument (Maybe [ExpenseField])

Any information found outside of a table by Amazon Textract.

ExpenseField

data ExpenseField

Breakdown of detected information, seperated into the catagories Type, LableDetection, and ValueDetection

See: newExpenseField smart constructor.

Constructors

ExpenseField'

Fields

labelDetection :: Maybe ExpenseDetection

The explicitly stated label of a detected element.

valueDetection :: Maybe ExpenseDetection

The value of a detected element. Present in explicit and implicit elements.

pageNumber :: Maybe Natural

The page number the value was detected on.

type' :: Maybe ExpenseType

The implied label of a detected element. Present alongside LabelDetection for explicit elements.

newExpenseField :: ExpenseField

Create a value of ExpenseField with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:labelDetection:ExpenseField', expenseField_labelDetection - The explicitly stated label of a detected element.

$sel:valueDetection:ExpenseField', expenseField_valueDetection - The value of a detected element. Present in explicit and implicit elements.

$sel:pageNumber:ExpenseField', expenseField_pageNumber - The page number the value was detected on.

$sel:type':ExpenseField', expenseField_type - The implied label of a detected element. Present alongside LabelDetection for explicit elements.

expenseField_labelDetection :: Lens' ExpenseField (Maybe ExpenseDetection)

The explicitly stated label of a detected element.

expenseField_valueDetection :: Lens' ExpenseField (Maybe ExpenseDetection)

The value of a detected element. Present in explicit and implicit elements.

expenseField_pageNumber :: Lens' ExpenseField (Maybe Natural)

The page number the value was detected on.

expenseField_type :: Lens' ExpenseField (Maybe ExpenseType)

The implied label of a detected element. Present alongside LabelDetection for explicit elements.

ExpenseType

data ExpenseType

An object used to store information about the Type detected by Amazon Textract.

See: newExpenseType smart constructor.

Constructors

ExpenseType'

Fields

confidence :: Maybe Double

The confidence of accuracy, as a percentage.

text :: Maybe Text

The word or line of text detected by Amazon Textract.

newExpenseType :: ExpenseType

Create a value of ExpenseType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:confidence:ExpenseType', expenseType_confidence - The confidence of accuracy, as a percentage.

$sel:text:ExpenseType', expenseType_text - The word or line of text detected by Amazon Textract.

expenseType_confidence :: Lens' ExpenseType (Maybe Double)

The confidence of accuracy, as a percentage.

expenseType_text :: Lens' ExpenseType (Maybe Text)

The word or line of text detected by Amazon Textract.

Geometry

data Geometry

Information about where the following items are located on a document page: detected page, text, key-value pairs, tables, table cells, and selection elements.

See: newGeometry smart constructor.

Constructors

Geometry'

Fields

polygon :: Maybe [Point]

Within the bounding box, a fine-grained polygon around the recognized item.

boundingBox :: Maybe BoundingBox

An axis-aligned coarse representation of the location of the recognized item on the document page.

newGeometry :: Geometry

Create a value of Geometry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:polygon:Geometry', geometry_polygon - Within the bounding box, a fine-grained polygon around the recognized item.

$sel:boundingBox:Geometry', geometry_boundingBox - An axis-aligned coarse representation of the location of the recognized item on the document page.

geometry_polygon :: Lens' Geometry (Maybe [Point])

Within the bounding box, a fine-grained polygon around the recognized item.

geometry_boundingBox :: Lens' Geometry (Maybe BoundingBox)

An axis-aligned coarse representation of the location of the recognized item on the document page.

HumanLoopActivationOutput

data HumanLoopActivationOutput

Shows the results of the human in the loop evaluation. If there is no HumanLoopArn, the input did not trigger human review.

See: newHumanLoopActivationOutput smart constructor.

Constructors

HumanLoopActivationOutput'

Fields

humanLoopActivationReasons :: Maybe (NonEmpty Text)

Shows if and why human review was needed.

humanLoopArn :: Maybe Text

The Amazon Resource Name (ARN) of the HumanLoop created.

humanLoopActivationConditionsEvaluationResults :: Maybe Text

Shows the result of condition evaluations, including those conditions which activated a human review.

newHumanLoopActivationOutput :: HumanLoopActivationOutput

Create a value of HumanLoopActivationOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:humanLoopActivationReasons:HumanLoopActivationOutput', humanLoopActivationOutput_humanLoopActivationReasons - Shows if and why human review was needed.

$sel:humanLoopArn:HumanLoopActivationOutput', humanLoopActivationOutput_humanLoopArn - The Amazon Resource Name (ARN) of the HumanLoop created.

$sel:humanLoopActivationConditionsEvaluationResults:HumanLoopActivationOutput', humanLoopActivationOutput_humanLoopActivationConditionsEvaluationResults - Shows the result of condition evaluations, including those conditions which activated a human review.

humanLoopActivationOutput_humanLoopActivationReasons :: Lens' HumanLoopActivationOutput (Maybe (NonEmpty Text))

Shows if and why human review was needed.

humanLoopActivationOutput_humanLoopArn :: Lens' HumanLoopActivationOutput (Maybe Text)

The Amazon Resource Name (ARN) of the HumanLoop created.

humanLoopActivationOutput_humanLoopActivationConditionsEvaluationResults :: Lens' HumanLoopActivationOutput (Maybe Text)

Shows the result of condition evaluations, including those conditions which activated a human review.

HumanLoopConfig

data HumanLoopConfig

Sets up the human review workflow the document will be sent to if one of the conditions is met. You can also set certain attributes of the image before review.

See: newHumanLoopConfig smart constructor.

Constructors

HumanLoopConfig'

Fields

dataAttributes :: Maybe HumanLoopDataAttributes

Sets attributes of the input data.

humanLoopName :: Text

The name of the human workflow used for this image. This should be kept unique within a region.

flowDefinitionArn :: Text

The Amazon Resource Name (ARN) of the flow definition.

newHumanLoopConfig

Arguments

:: Text
$sel:humanLoopName:HumanLoopConfig'

-> Text
$sel:flowDefinitionArn:HumanLoopConfig'

-> HumanLoopConfig

Create a value of HumanLoopConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataAttributes:HumanLoopConfig', humanLoopConfig_dataAttributes - Sets attributes of the input data.

$sel:humanLoopName:HumanLoopConfig', humanLoopConfig_humanLoopName - The name of the human workflow used for this image. This should be kept unique within a region.

$sel:flowDefinitionArn:HumanLoopConfig', humanLoopConfig_flowDefinitionArn - The Amazon Resource Name (ARN) of the flow definition.

humanLoopConfig_dataAttributes :: Lens' HumanLoopConfig (Maybe HumanLoopDataAttributes)

Sets attributes of the input data.

humanLoopConfig_humanLoopName :: Lens' HumanLoopConfig Text

The name of the human workflow used for this image. This should be kept unique within a region.

humanLoopConfig_flowDefinitionArn :: Lens' HumanLoopConfig Text

The Amazon Resource Name (ARN) of the flow definition.

HumanLoopDataAttributes

data HumanLoopDataAttributes

Allows you to set attributes of the image. Currently, you can declare an image as free of personally identifiable information and adult content.

See: newHumanLoopDataAttributes smart constructor.

Constructors

HumanLoopDataAttributes'

Fields

contentClassifiers :: Maybe [ContentClassifier]

Sets whether the input image is free of personally identifiable information or adult content.

newHumanLoopDataAttributes :: HumanLoopDataAttributes

Create a value of HumanLoopDataAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentClassifiers:HumanLoopDataAttributes', humanLoopDataAttributes_contentClassifiers - Sets whether the input image is free of personally identifiable information or adult content.

humanLoopDataAttributes_contentClassifiers :: Lens' HumanLoopDataAttributes (Maybe [ContentClassifier])

Sets whether the input image is free of personally identifiable information or adult content.

LineItemFields

data LineItemFields

A structure that holds information about the different lines found in a document's tables.

See: newLineItemFields smart constructor.

Constructors

LineItemFields'

Fields

lineItemExpenseFields :: Maybe [ExpenseField]

ExpenseFields used to show information from detected lines on a table.

newLineItemFields :: LineItemFields

Create a value of LineItemFields with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lineItemExpenseFields:LineItemFields', lineItemFields_lineItemExpenseFields - ExpenseFields used to show information from detected lines on a table.

lineItemFields_lineItemExpenseFields :: Lens' LineItemFields (Maybe [ExpenseField])

ExpenseFields used to show information from detected lines on a table.

LineItemGroup

data LineItemGroup

A grouping of tables which contain LineItems, with each table identified by the table's LineItemGroupIndex.

See: newLineItemGroup smart constructor.

Constructors

LineItemGroup'

Fields

lineItems :: Maybe [LineItemFields]

The breakdown of information on a particular line of a table.

lineItemGroupIndex :: Maybe Natural

The number used to identify a specific table in a document. The first table encountered will have a LineItemGroupIndex of 1, the second 2, etc.

newLineItemGroup :: LineItemGroup

Create a value of LineItemGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lineItems:LineItemGroup', lineItemGroup_lineItems - The breakdown of information on a particular line of a table.

$sel:lineItemGroupIndex:LineItemGroup', lineItemGroup_lineItemGroupIndex - The number used to identify a specific table in a document. The first table encountered will have a LineItemGroupIndex of 1, the second 2, etc.

lineItemGroup_lineItems :: Lens' LineItemGroup (Maybe [LineItemFields])

The breakdown of information on a particular line of a table.

lineItemGroup_lineItemGroupIndex :: Lens' LineItemGroup (Maybe Natural)

The number used to identify a specific table in a document. The first table encountered will have a LineItemGroupIndex of 1, the second 2, etc.

NotificationChannel

data NotificationChannel

The Amazon Simple Notification Service (Amazon SNS) topic to which Amazon Textract publishes the completion status of an asynchronous document operation, such as StartDocumentTextDetection.

See: newNotificationChannel smart constructor.

Constructors

NotificationChannel'

Fields

sNSTopicArn :: Text

The Amazon SNS topic that Amazon Textract posts the completion status to.

roleArn :: Text

The Amazon Resource Name (ARN) of an IAM role that gives Amazon Textract publishing permissions to the Amazon SNS topic.

newNotificationChannel

Arguments

:: Text
$sel:sNSTopicArn:NotificationChannel'

-> Text
$sel:roleArn:NotificationChannel'

-> NotificationChannel

Create a value of NotificationChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sNSTopicArn:NotificationChannel', notificationChannel_sNSTopicArn - The Amazon SNS topic that Amazon Textract posts the completion status to.

$sel:roleArn:NotificationChannel', notificationChannel_roleArn - The Amazon Resource Name (ARN) of an IAM role that gives Amazon Textract publishing permissions to the Amazon SNS topic.

notificationChannel_sNSTopicArn :: Lens' NotificationChannel Text

The Amazon SNS topic that Amazon Textract posts the completion status to.

notificationChannel_roleArn :: Lens' NotificationChannel Text

The Amazon Resource Name (ARN) of an IAM role that gives Amazon Textract publishing permissions to the Amazon SNS topic.

OutputConfig

data OutputConfig

Sets whether or not your output will go to a user created bucket. Used to set the name of the bucket, and the prefix on the output file.

OutputConfig is an optional parameter which lets you adjust where your output will be placed. By default, Amazon Textract will store the results internally and can only be accessed by the Get API operations. With OutputConfig enabled, you can set the name of the bucket the output will be sent to and the file prefix of the results where you can download your results. Additionally, you can set the KMSKeyID parameter to a customer master key (CMK) to encrypt your output. Without this parameter set Amazon Textract will encrypt server-side using the AWS managed CMK for Amazon S3.

Decryption of Customer Content is necessary for processing of the documents by Amazon Textract. If your account is opted out under an AI services opt out policy then all unencrypted Customer Content is immediately and permanently deleted after the Customer Content has been processed by the service. No copy of of the output is retained by Amazon Textract. For information about how to opt out, see Managing AI services opt-out policy.

For more information on data privacy, see the Data Privacy FAQ.

See: newOutputConfig smart constructor.

Constructors

OutputConfig'

Fields

s3Prefix :: Maybe Text

The prefix of the object key that the output will be saved to. When not enabled, the prefix will be “textract_output".

s3Bucket :: Text

The name of the bucket your output will go to.

newOutputConfig

Arguments

:: Text
$sel:s3Bucket:OutputConfig'

-> OutputConfig

Create a value of OutputConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Prefix:OutputConfig', outputConfig_s3Prefix - The prefix of the object key that the output will be saved to. When not enabled, the prefix will be “textract_output".

$sel:s3Bucket:OutputConfig', outputConfig_s3Bucket - The name of the bucket your output will go to.

outputConfig_s3Prefix :: Lens' OutputConfig (Maybe Text)

The prefix of the object key that the output will be saved to. When not enabled, the prefix will be “textract_output".

outputConfig_s3Bucket :: Lens' OutputConfig Text

The name of the bucket your output will go to.

Point

data Point

The X and Y coordinates of a point on a document page. The X and Y values that are returned are ratios of the overall document page size. For example, if the input document is 700 x 200 and the operation returns X=0.5 and Y=0.25, then the point is at the (350,50) pixel coordinate on the document page.

An array of Point objects, Polygon, is returned by DetectDocumentText. Polygon represents a fine-grained polygon around detected text. For more information, see Geometry in the Amazon Textract Developer Guide.

See: newPoint smart constructor.

Constructors

Point'

Fields

y :: Maybe Double

The value of the Y coordinate for a point on a Polygon.

x :: Maybe Double

The value of the X coordinate for a point on a Polygon.

newPoint :: Point

Create a value of Point with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:y:Point', point_y - The value of the Y coordinate for a point on a Polygon.

$sel:x:Point', point_x - The value of the X coordinate for a point on a Polygon.

point_y :: Lens' Point (Maybe Double)

The value of the Y coordinate for a point on a Polygon.

point_x :: Lens' Point (Maybe Double)

The value of the X coordinate for a point on a Polygon.

Relationship

data Relationship

Information about how blocks are related to each other. A Block object contains 0 or more Relation objects in a list, Relationships. For more information, see Block.

The Type element provides the type of the relationship for all blocks in the IDs array.

See: newRelationship smart constructor.

Constructors

Relationship'

Fields

ids :: Maybe [Text]

An array of IDs for related blocks. You can get the type of the relationship from the Type element.

type' :: Maybe RelationshipType

The type of relationship that the blocks in the IDs array have with the current block. The relationship can be VALUE or CHILD. A relationship of type VALUE is a list that contains the ID of the VALUE block that's associated with the KEY of a key-value pair. A relationship of type CHILD is a list of IDs that identify WORD blocks in the case of lines Cell blocks in the case of Tables, and WORD blocks in the case of Selection Elements.

newRelationship :: Relationship

Create a value of Relationship with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ids:Relationship', relationship_ids - An array of IDs for related blocks. You can get the type of the relationship from the Type element.

$sel:type':Relationship', relationship_type - The type of relationship that the blocks in the IDs array have with the current block. The relationship can be VALUE or CHILD. A relationship of type VALUE is a list that contains the ID of the VALUE block that's associated with the KEY of a key-value pair. A relationship of type CHILD is a list of IDs that identify WORD blocks in the case of lines Cell blocks in the case of Tables, and WORD blocks in the case of Selection Elements.

relationship_ids :: Lens' Relationship (Maybe [Text])

An array of IDs for related blocks. You can get the type of the relationship from the Type element.

relationship_type :: Lens' Relationship (Maybe RelationshipType)

The type of relationship that the blocks in the IDs array have with the current block. The relationship can be VALUE or CHILD. A relationship of type VALUE is a list that contains the ID of the VALUE block that's associated with the KEY of a key-value pair. A relationship of type CHILD is a list of IDs that identify WORD blocks in the case of lines Cell blocks in the case of Tables, and WORD blocks in the case of Selection Elements.

S3Object

data S3Object

The S3 bucket name and file name that identifies the document.

The AWS Region for the S3 bucket that contains the document must match the Region that you use for Amazon Textract operations.

For Amazon Textract to process a file in an S3 bucket, the user must have permission to access the S3 bucket and file.

See: newS3Object smart constructor.

Constructors

S3Object'

Fields

version :: Maybe Text

If the bucket has versioning enabled, you can specify the object version.

name :: Maybe Text

The file name of the input document. Synchronous operations can use image files that are in JPEG or PNG format. Asynchronous operations also support PDF format files.

bucket :: Maybe Text

The name of the S3 bucket. Note that the # character is not valid in the file name.

newS3Object :: S3Object

Create a value of S3Object with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:version:S3Object', s3Object_version - If the bucket has versioning enabled, you can specify the object version.

$sel:name:S3Object', s3Object_name - The file name of the input document. Synchronous operations can use image files that are in JPEG or PNG format. Asynchronous operations also support PDF format files.

$sel:bucket:S3Object', s3Object_bucket - The name of the S3 bucket. Note that the # character is not valid in the file name.

s3Object_version :: Lens' S3Object (Maybe Text)

If the bucket has versioning enabled, you can specify the object version.

s3Object_name :: Lens' S3Object (Maybe Text)

The file name of the input document. Synchronous operations can use image files that are in JPEG or PNG format. Asynchronous operations also support PDF format files.

s3Object_bucket :: Lens' S3Object (Maybe Text)

The name of the S3 bucket. Note that the # character is not valid in the file name.

Warning

data Warning

A warning about an issue that occurred during asynchronous text analysis (StartDocumentAnalysis) or asynchronous document text detection (StartDocumentTextDetection).

See: newWarning smart constructor.

Constructors

Warning'

Fields

pages :: Maybe [Natural]

A list of the pages that the warning applies to.

errorCode :: Maybe Text

The error code for the warning.

newWarning :: Warning

Create a value of Warning with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pages:Warning', warning_pages - A list of the pages that the warning applies to.

$sel:errorCode:Warning', warning_errorCode - The error code for the warning.

warning_pages :: Lens' Warning (Maybe [Natural])

A list of the pages that the warning applies to.

warning_errorCode :: Lens' Warning (Maybe Text)

The error code for the warning.