Network.AWS.Textract.Types.ExpenseField

Description

Documentation

data ExpenseField

Breakdown of detected information, seperated into the catagories Type, LableDetection, and ValueDetection

See: newExpenseField smart constructor.

Constructors

ExpenseField'

Fields

labelDetection :: Maybe ExpenseDetection

The explicitly stated label of a detected element.

valueDetection :: Maybe ExpenseDetection

The value of a detected element. Present in explicit and implicit elements.

type' :: Maybe ExpenseType

The implied label of a detected element. Present alongside LabelDetection for explicit elements.

pageNumber :: Maybe Natural

The page number the value was detected on.

newExpenseField :: ExpenseField

Create a value of ExpenseField with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:labelDetection:ExpenseField', expenseField_labelDetection - The explicitly stated label of a detected element.

$sel:valueDetection:ExpenseField', expenseField_valueDetection - The value of a detected element. Present in explicit and implicit elements.

$sel:type':ExpenseField', expenseField_type - The implied label of a detected element. Present alongside LabelDetection for explicit elements.

$sel:pageNumber:ExpenseField', expenseField_pageNumber - The page number the value was detected on.

expenseField_labelDetection :: Lens' ExpenseField (Maybe ExpenseDetection)

The explicitly stated label of a detected element.

expenseField_valueDetection :: Lens' ExpenseField (Maybe ExpenseDetection)

The value of a detected element. Present in explicit and implicit elements.

expenseField_type :: Lens' ExpenseField (Maybe ExpenseType)

The implied label of a detected element. Present alongside LabelDetection for explicit elements.

expenseField_pageNumber :: Lens' ExpenseField (Maybe Natural)

The page number the value was detected on.