Network.AWS.Transcribe.CreateCallAnalyticsCategory

Description

Creates an analytics category. Amazon Transcribe applies the conditions specified by your analytics categories to your call analytics jobs. For each analytics category, you specify one or more rules. For example, you can specify a rule that the customer sentiment was neutral or negative within that category. If you start a call analytics job, Amazon Transcribe applies the category to the analytics job that you've specified.

Creating a Request

data CreateCallAnalyticsCategory

See: newCreateCallAnalyticsCategory smart constructor.

Constructors

CreateCallAnalyticsCategory'

Fields

categoryName :: Text

The name that you choose for your category when you create it.

rules :: NonEmpty Rule

To create a category, you must specify between 1 and 20 rules. For each rule, you specify a filter to be applied to the attributes of the call. For example, you can specify a sentiment filter to detect if the customer's sentiment was negative or neutral.

newCreateCallAnalyticsCategory

Arguments

:: Text
$sel:categoryName:CreateCallAnalyticsCategory'

-> NonEmpty Rule
$sel:rules:CreateCallAnalyticsCategory'

-> CreateCallAnalyticsCategory

Create a value of CreateCallAnalyticsCategory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:categoryName:CreateCallAnalyticsCategory', createCallAnalyticsCategory_categoryName - The name that you choose for your category when you create it.

$sel:rules:CreateCallAnalyticsCategory', createCallAnalyticsCategory_rules - To create a category, you must specify between 1 and 20 rules. For each rule, you specify a filter to be applied to the attributes of the call. For example, you can specify a sentiment filter to detect if the customer's sentiment was negative or neutral.

Request Lenses

createCallAnalyticsCategory_categoryName :: Lens' CreateCallAnalyticsCategory Text

The name that you choose for your category when you create it.

createCallAnalyticsCategory_rules :: Lens' CreateCallAnalyticsCategory (NonEmpty Rule)

To create a category, you must specify between 1 and 20 rules. For each rule, you specify a filter to be applied to the attributes of the call. For example, you can specify a sentiment filter to detect if the customer's sentiment was negative or neutral.

Destructuring the Response

data CreateCallAnalyticsCategoryResponse

See: newCreateCallAnalyticsCategoryResponse smart constructor.

Constructors

CreateCallAnalyticsCategoryResponse'

Fields

categoryProperties :: Maybe CategoryProperties

The rules and associated metadata used to create a category.

httpStatus :: Int

The response's http status code.

newCreateCallAnalyticsCategoryResponse

Arguments

:: Int
$sel:httpStatus:CreateCallAnalyticsCategoryResponse'

-> CreateCallAnalyticsCategoryResponse

Create a value of CreateCallAnalyticsCategoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:categoryProperties:CreateCallAnalyticsCategoryResponse', createCallAnalyticsCategoryResponse_categoryProperties - The rules and associated metadata used to create a category.

$sel:httpStatus:CreateCallAnalyticsCategoryResponse', createCallAnalyticsCategoryResponse_httpStatus - The response's http status code.

Response Lenses

createCallAnalyticsCategoryResponse_categoryProperties :: Lens' CreateCallAnalyticsCategoryResponse (Maybe CategoryProperties)

The rules and associated metadata used to create a category.

createCallAnalyticsCategoryResponse_httpStatus :: Lens' CreateCallAnalyticsCategoryResponse Int

The response's http status code.