Network.AWS.Transcribe.CreateLanguageModel

Description

Creates a new custom language model. Use Amazon S3 prefixes to provide the location of your input files. The time it takes to create your model depends on the size of your training data.

Creating a Request

data CreateLanguageModel

See: newCreateLanguageModel smart constructor.

Constructors

CreateLanguageModel'

Fields

tags :: Maybe (NonEmpty Tag)

Adds one or more tags, each in the form of a key:value pair, to a new language model at the time you create this new model.

languageCode :: CLMLanguageCode

The language of the input text you're using to train your custom language model.

baseModelName :: BaseModelName

The Amazon Transcribe standard language model, or base model used to create your custom language model.

If you want to use your custom language model to transcribe audio with a sample rate of 16,000 Hz or greater, choose Wideband.

If you want to use your custom language model to transcribe audio with a sample rate that is less than 16,000 Hz, choose Narrowband.

modelName :: Text

The name you choose for your custom language model when you create it.

inputDataConfig :: InputDataConfig

Contains the data access role and the Amazon S3 prefixes to read the required input files to create a custom language model.

newCreateLanguageModel

Arguments

:: CLMLanguageCode
$sel:languageCode:CreateLanguageModel'

-> BaseModelName
$sel:baseModelName:CreateLanguageModel'

-> Text
$sel:modelName:CreateLanguageModel'

-> InputDataConfig
$sel:inputDataConfig:CreateLanguageModel'

-> CreateLanguageModel

Create a value of CreateLanguageModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateLanguageModel', createLanguageModel_tags - Adds one or more tags, each in the form of a key:value pair, to a new language model at the time you create this new model.

$sel:languageCode:CreateLanguageModel', createLanguageModel_languageCode - The language of the input text you're using to train your custom language model.

$sel:baseModelName:CreateLanguageModel', createLanguageModel_baseModelName - The Amazon Transcribe standard language model, or base model used to create your custom language model.

If you want to use your custom language model to transcribe audio with a sample rate of 16,000 Hz or greater, choose Wideband.

If you want to use your custom language model to transcribe audio with a sample rate that is less than 16,000 Hz, choose Narrowband.

$sel:modelName:CreateLanguageModel', createLanguageModel_modelName - The name you choose for your custom language model when you create it.

$sel:inputDataConfig:CreateLanguageModel', createLanguageModel_inputDataConfig - Contains the data access role and the Amazon S3 prefixes to read the required input files to create a custom language model.

Request Lenses

createLanguageModel_tags :: Lens' CreateLanguageModel (Maybe (NonEmpty Tag))

Adds one or more tags, each in the form of a key:value pair, to a new language model at the time you create this new model.

createLanguageModel_languageCode :: Lens' CreateLanguageModel CLMLanguageCode

The language of the input text you're using to train your custom language model.

createLanguageModel_baseModelName :: Lens' CreateLanguageModel BaseModelName

The Amazon Transcribe standard language model, or base model used to create your custom language model.

If you want to use your custom language model to transcribe audio with a sample rate of 16,000 Hz or greater, choose Wideband.

If you want to use your custom language model to transcribe audio with a sample rate that is less than 16,000 Hz, choose Narrowband.

createLanguageModel_modelName :: Lens' CreateLanguageModel Text

The name you choose for your custom language model when you create it.

createLanguageModel_inputDataConfig :: Lens' CreateLanguageModel InputDataConfig

Contains the data access role and the Amazon S3 prefixes to read the required input files to create a custom language model.

Destructuring the Response

data CreateLanguageModelResponse

See: newCreateLanguageModelResponse smart constructor.

Constructors

CreateLanguageModelResponse'

Fields

languageCode :: Maybe CLMLanguageCode

The language code of the text you've used to create a custom language model.

inputDataConfig :: Maybe InputDataConfig

The data access role and Amazon S3 prefixes you've chosen to create your custom language model.

modelStatus :: Maybe ModelStatus

The status of the custom language model. When the status is COMPLETED the model is ready to use.

modelName :: Maybe Text

The name you've chosen for your custom language model.

baseModelName :: Maybe BaseModelName

The Amazon Transcribe standard language model, or base model you've used to create a custom language model.

httpStatus :: Int

The response's http status code.

newCreateLanguageModelResponse

Arguments

:: Int
$sel:httpStatus:CreateLanguageModelResponse'

-> CreateLanguageModelResponse

Create a value of CreateLanguageModelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:CreateLanguageModel', createLanguageModelResponse_languageCode - The language code of the text you've used to create a custom language model.

$sel:inputDataConfig:CreateLanguageModel', createLanguageModelResponse_inputDataConfig - The data access role and Amazon S3 prefixes you've chosen to create your custom language model.

$sel:modelStatus:CreateLanguageModelResponse', createLanguageModelResponse_modelStatus - The status of the custom language model. When the status is COMPLETED the model is ready to use.

$sel:modelName:CreateLanguageModel', createLanguageModelResponse_modelName - The name you've chosen for your custom language model.

$sel:baseModelName:CreateLanguageModel', createLanguageModelResponse_baseModelName - The Amazon Transcribe standard language model, or base model you've used to create a custom language model.

$sel:httpStatus:CreateLanguageModelResponse', createLanguageModelResponse_httpStatus - The response's http status code.

Response Lenses

createLanguageModelResponse_languageCode :: Lens' CreateLanguageModelResponse (Maybe CLMLanguageCode)

The language code of the text you've used to create a custom language model.

createLanguageModelResponse_inputDataConfig :: Lens' CreateLanguageModelResponse (Maybe InputDataConfig)

The data access role and Amazon S3 prefixes you've chosen to create your custom language model.

createLanguageModelResponse_modelStatus :: Lens' CreateLanguageModelResponse (Maybe ModelStatus)

The status of the custom language model. When the status is COMPLETED the model is ready to use.

createLanguageModelResponse_modelName :: Lens' CreateLanguageModelResponse (Maybe Text)

The name you've chosen for your custom language model.

createLanguageModelResponse_baseModelName :: Lens' CreateLanguageModelResponse (Maybe BaseModelName)

The Amazon Transcribe standard language model, or base model you've used to create a custom language model.

createLanguageModelResponse_httpStatus :: Lens' CreateLanguageModelResponse Int

The response's http status code.