Network.AWS.Transcribe.CreateMedicalVocabulary

Description

Creates a new custom vocabulary that you can use to modify how Amazon Transcribe Medical transcribes your audio file.

Creating a Request

data CreateMedicalVocabulary

See: newCreateMedicalVocabulary smart constructor.

Constructors

CreateMedicalVocabulary'

Fields

tags :: Maybe (NonEmpty Tag)

Adds one or more tags, each in the form of a key:value pair, to a new medical vocabulary at the time you create this new vocabulary.

vocabularyName :: Text

The name of the custom vocabulary. This case-sensitive name must be unique within an Amazon Web Services account. If you try to create a vocabulary with the same name as a previous vocabulary, you get a ConflictException error.

languageCode :: LanguageCode

The language code for the language used for the entries in your custom vocabulary. The language code of your custom vocabulary must match the language code of your transcription job. US English (en-US) is the only language code available for Amazon Transcribe Medical.

vocabularyFileUri :: Text

The location in Amazon S3 of the text file you use to define your custom vocabulary. The URI must be in the same Amazon Web Services Region as the resource that you're calling. Enter information about your VocabularyFileUri in the following format:

 https://s3.<aws-region>.amazonaws.com/<bucket-name>/<keyprefix>/<objectkey>

The following is an example URI for a vocabulary file that is stored in Amazon S3:

https://s3.us-east-1.amazonaws.com/AWSDOC-EXAMPLE-BUCKET/vocab.txt

For more information about Amazon S3 object names, see Object Keys in the Amazon S3 Developer Guide.

For more information about custom vocabularies, see Medical Custom Vocabularies.

newCreateMedicalVocabulary

Arguments

:: Text
$sel:vocabularyName:CreateMedicalVocabulary'

-> LanguageCode
$sel:languageCode:CreateMedicalVocabulary'

-> Text
$sel:vocabularyFileUri:CreateMedicalVocabulary'

-> CreateMedicalVocabulary

Create a value of CreateMedicalVocabulary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateMedicalVocabulary', createMedicalVocabulary_tags - Adds one or more tags, each in the form of a key:value pair, to a new medical vocabulary at the time you create this new vocabulary.

$sel:vocabularyName:CreateMedicalVocabulary', createMedicalVocabulary_vocabularyName - The name of the custom vocabulary. This case-sensitive name must be unique within an Amazon Web Services account. If you try to create a vocabulary with the same name as a previous vocabulary, you get a ConflictException error.

$sel:languageCode:CreateMedicalVocabulary', createMedicalVocabulary_languageCode - The language code for the language used for the entries in your custom vocabulary. The language code of your custom vocabulary must match the language code of your transcription job. US English (en-US) is the only language code available for Amazon Transcribe Medical.

$sel:vocabularyFileUri:CreateMedicalVocabulary', createMedicalVocabulary_vocabularyFileUri - The location in Amazon S3 of the text file you use to define your custom vocabulary. The URI must be in the same Amazon Web Services Region as the resource that you're calling. Enter information about your VocabularyFileUri in the following format:

 https://s3.<aws-region>.amazonaws.com/<bucket-name>/<keyprefix>/<objectkey>

The following is an example URI for a vocabulary file that is stored in Amazon S3:

https://s3.us-east-1.amazonaws.com/AWSDOC-EXAMPLE-BUCKET/vocab.txt

For more information about Amazon S3 object names, see Object Keys in the Amazon S3 Developer Guide.

For more information about custom vocabularies, see Medical Custom Vocabularies.

Request Lenses

createMedicalVocabulary_tags :: Lens' CreateMedicalVocabulary (Maybe (NonEmpty Tag))

Adds one or more tags, each in the form of a key:value pair, to a new medical vocabulary at the time you create this new vocabulary.

createMedicalVocabulary_vocabularyName :: Lens' CreateMedicalVocabulary Text

The name of the custom vocabulary. This case-sensitive name must be unique within an Amazon Web Services account. If you try to create a vocabulary with the same name as a previous vocabulary, you get a ConflictException error.

createMedicalVocabulary_languageCode :: Lens' CreateMedicalVocabulary LanguageCode

The language code for the language used for the entries in your custom vocabulary. The language code of your custom vocabulary must match the language code of your transcription job. US English (en-US) is the only language code available for Amazon Transcribe Medical.

createMedicalVocabulary_vocabularyFileUri :: Lens' CreateMedicalVocabulary Text

The location in Amazon S3 of the text file you use to define your custom vocabulary. The URI must be in the same Amazon Web Services Region as the resource that you're calling. Enter information about your VocabularyFileUri in the following format:

 https://s3.<aws-region>.amazonaws.com/<bucket-name>/<keyprefix>/<objectkey>

The following is an example URI for a vocabulary file that is stored in Amazon S3:

https://s3.us-east-1.amazonaws.com/AWSDOC-EXAMPLE-BUCKET/vocab.txt

For more information about Amazon S3 object names, see Object Keys in the Amazon S3 Developer Guide.

For more information about custom vocabularies, see Medical Custom Vocabularies.

Destructuring the Response

data CreateMedicalVocabularyResponse

See: newCreateMedicalVocabularyResponse smart constructor.

Constructors

CreateMedicalVocabularyResponse'

Fields

failureReason :: Maybe Text

If the VocabularyState field is FAILED, this field contains information about why the job failed.

languageCode :: Maybe LanguageCode

The language code for the entries in your custom vocabulary. US English (en-US) is the only valid language code for Amazon Transcribe Medical.

vocabularyName :: Maybe Text

The name of the vocabulary. The name must be unique within an Amazon Web Services account and is case sensitive.

lastModifiedTime :: Maybe POSIX

The date and time that you created the vocabulary.

vocabularyState :: Maybe VocabularyState

The processing state of your custom vocabulary in Amazon Transcribe Medical. If the state is READY, you can use the vocabulary in a StartMedicalTranscriptionJob request.

httpStatus :: Int

The response's http status code.

newCreateMedicalVocabularyResponse

Arguments

:: Int
$sel:httpStatus:CreateMedicalVocabularyResponse'

-> CreateMedicalVocabularyResponse

Create a value of CreateMedicalVocabularyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:CreateMedicalVocabularyResponse', createMedicalVocabularyResponse_failureReason - If the VocabularyState field is FAILED, this field contains information about why the job failed.

$sel:languageCode:CreateMedicalVocabulary', createMedicalVocabularyResponse_languageCode - The language code for the entries in your custom vocabulary. US English (en-US) is the only valid language code for Amazon Transcribe Medical.

$sel:vocabularyName:CreateMedicalVocabulary', createMedicalVocabularyResponse_vocabularyName - The name of the vocabulary. The name must be unique within an Amazon Web Services account and is case sensitive.

$sel:lastModifiedTime:CreateMedicalVocabularyResponse', createMedicalVocabularyResponse_lastModifiedTime - The date and time that you created the vocabulary.

$sel:vocabularyState:CreateMedicalVocabularyResponse', createMedicalVocabularyResponse_vocabularyState - The processing state of your custom vocabulary in Amazon Transcribe Medical. If the state is READY, you can use the vocabulary in a StartMedicalTranscriptionJob request.

$sel:httpStatus:CreateMedicalVocabularyResponse', createMedicalVocabularyResponse_httpStatus - The response's http status code.

Response Lenses

createMedicalVocabularyResponse_failureReason :: Lens' CreateMedicalVocabularyResponse (Maybe Text)

If the VocabularyState field is FAILED, this field contains information about why the job failed.

createMedicalVocabularyResponse_languageCode :: Lens' CreateMedicalVocabularyResponse (Maybe LanguageCode)

The language code for the entries in your custom vocabulary. US English (en-US) is the only valid language code for Amazon Transcribe Medical.

createMedicalVocabularyResponse_vocabularyName :: Lens' CreateMedicalVocabularyResponse (Maybe Text)

The name of the vocabulary. The name must be unique within an Amazon Web Services account and is case sensitive.

createMedicalVocabularyResponse_lastModifiedTime :: Lens' CreateMedicalVocabularyResponse (Maybe UTCTime)

The date and time that you created the vocabulary.

createMedicalVocabularyResponse_vocabularyState :: Lens' CreateMedicalVocabularyResponse (Maybe VocabularyState)

The processing state of your custom vocabulary in Amazon Transcribe Medical. If the state is READY, you can use the vocabulary in a StartMedicalTranscriptionJob request.

createMedicalVocabularyResponse_httpStatus :: Lens' CreateMedicalVocabularyResponse Int

The response's http status code.