Network.AWS.Transcribe.CreateVocabulary

Description

Creates a new custom vocabulary that you can use to change the way Amazon Transcribe handles transcription of an audio file.

Creating a Request

data CreateVocabulary

See: newCreateVocabulary smart constructor.

Constructors

CreateVocabulary'

Fields

phrases :: Maybe [Text]

An array of strings that contains the vocabulary entries.

tags :: Maybe (NonEmpty Tag)

Adds one or more tags, each in the form of a key:value pair, to a new Amazon Transcribe vocabulary at the time you create this new vocabulary.

vocabularyFileUri :: Maybe Text

The S3 location of the text file that contains the definition of the custom vocabulary. The URI must be in the same region as the API endpoint that you are calling. The general form is:

For more information about S3 object names, see Object Keys in the Amazon S3 Developer Guide.

For more information about custom vocabularies, see Custom vocabularies.

vocabularyName :: Text

The name of the vocabulary. The name must be unique within an Amazon Web Services account. The name is case sensitive. If you try to create a vocabulary with the same name as a previous vocabulary you will receive a ConflictException error.

languageCode :: LanguageCode

The language code of the vocabulary entries. For a list of languages and their corresponding language codes, see transcribe-whatis.

newCreateVocabulary

Arguments

:: Text
$sel:vocabularyName:CreateVocabulary'

-> LanguageCode
$sel:languageCode:CreateVocabulary'

-> CreateVocabulary

Create a value of CreateVocabulary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phrases:CreateVocabulary', createVocabulary_phrases - An array of strings that contains the vocabulary entries.

$sel:tags:CreateVocabulary', createVocabulary_tags - Adds one or more tags, each in the form of a key:value pair, to a new Amazon Transcribe vocabulary at the time you create this new vocabulary.

$sel:vocabularyFileUri:CreateVocabulary', createVocabulary_vocabularyFileUri - The S3 location of the text file that contains the definition of the custom vocabulary. The URI must be in the same region as the API endpoint that you are calling. The general form is:

For more information about S3 object names, see Object Keys in the Amazon S3 Developer Guide.

For more information about custom vocabularies, see Custom vocabularies.

$sel:vocabularyName:CreateVocabulary', createVocabulary_vocabularyName - The name of the vocabulary. The name must be unique within an Amazon Web Services account. The name is case sensitive. If you try to create a vocabulary with the same name as a previous vocabulary you will receive a ConflictException error.

$sel:languageCode:CreateVocabulary', createVocabulary_languageCode - The language code of the vocabulary entries. For a list of languages and their corresponding language codes, see transcribe-whatis.

Request Lenses

createVocabulary_phrases :: Lens' CreateVocabulary (Maybe [Text])

An array of strings that contains the vocabulary entries.

createVocabulary_tags :: Lens' CreateVocabulary (Maybe (NonEmpty Tag))

Adds one or more tags, each in the form of a key:value pair, to a new Amazon Transcribe vocabulary at the time you create this new vocabulary.

createVocabulary_vocabularyFileUri :: Lens' CreateVocabulary (Maybe Text)

The S3 location of the text file that contains the definition of the custom vocabulary. The URI must be in the same region as the API endpoint that you are calling. The general form is:

For more information about S3 object names, see Object Keys in the Amazon S3 Developer Guide.

For more information about custom vocabularies, see Custom vocabularies.

createVocabulary_vocabularyName :: Lens' CreateVocabulary Text

The name of the vocabulary. The name must be unique within an Amazon Web Services account. The name is case sensitive. If you try to create a vocabulary with the same name as a previous vocabulary you will receive a ConflictException error.

createVocabulary_languageCode :: Lens' CreateVocabulary LanguageCode

The language code of the vocabulary entries. For a list of languages and their corresponding language codes, see transcribe-whatis.

Destructuring the Response

data CreateVocabularyResponse

See: newCreateVocabularyResponse smart constructor.

Constructors

CreateVocabularyResponse'

Fields

languageCode :: Maybe LanguageCode

The language code of the vocabulary entries.

failureReason :: Maybe Text

If the VocabularyState field is FAILED, this field contains information about why the job failed.

lastModifiedTime :: Maybe POSIX

The date and time that the vocabulary was created.

vocabularyName :: Maybe Text

The name of the vocabulary.

vocabularyState :: Maybe VocabularyState

The processing state of the vocabulary. When the VocabularyState field contains READY the vocabulary is ready to be used in a StartTranscriptionJob request.

httpStatus :: Int

The response's http status code.

newCreateVocabularyResponse

Arguments

:: Int
$sel:httpStatus:CreateVocabularyResponse'

-> CreateVocabularyResponse

Create a value of CreateVocabularyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:CreateVocabulary', createVocabularyResponse_languageCode - The language code of the vocabulary entries.

$sel:failureReason:CreateVocabularyResponse', createVocabularyResponse_failureReason - If the VocabularyState field is FAILED, this field contains information about why the job failed.

$sel:lastModifiedTime:CreateVocabularyResponse', createVocabularyResponse_lastModifiedTime - The date and time that the vocabulary was created.

$sel:vocabularyName:CreateVocabulary', createVocabularyResponse_vocabularyName - The name of the vocabulary.

$sel:vocabularyState:CreateVocabularyResponse', createVocabularyResponse_vocabularyState - The processing state of the vocabulary. When the VocabularyState field contains READY the vocabulary is ready to be used in a StartTranscriptionJob request.

$sel:httpStatus:CreateVocabularyResponse', createVocabularyResponse_httpStatus - The response's http status code.

Response Lenses

createVocabularyResponse_languageCode :: Lens' CreateVocabularyResponse (Maybe LanguageCode)

The language code of the vocabulary entries.

createVocabularyResponse_failureReason :: Lens' CreateVocabularyResponse (Maybe Text)

If the VocabularyState field is FAILED, this field contains information about why the job failed.

createVocabularyResponse_lastModifiedTime :: Lens' CreateVocabularyResponse (Maybe UTCTime)

The date and time that the vocabulary was created.

createVocabularyResponse_vocabularyName :: Lens' CreateVocabularyResponse (Maybe Text)

The name of the vocabulary.

createVocabularyResponse_vocabularyState :: Lens' CreateVocabularyResponse (Maybe VocabularyState)

The processing state of the vocabulary. When the VocabularyState field contains READY the vocabulary is ready to be used in a StartTranscriptionJob request.

createVocabularyResponse_httpStatus :: Lens' CreateVocabularyResponse Int

The response's http status code.