Network.AWS.Transcribe.GetCallAnalyticsJob

Description

Returns information about a call analytics job. To see the status of the job, check the CallAnalyticsJobStatus field. If the status is COMPLETED, the job is finished and you can find the results at the location specified in the TranscriptFileUri field. If you enable personally identifiable information (PII) redaction, the redacted transcript appears in the RedactedTranscriptFileUri field.

Creating a Request

data GetCallAnalyticsJob

See: newGetCallAnalyticsJob smart constructor.

Constructors

GetCallAnalyticsJob'

Fields

callAnalyticsJobName :: Text

The name of the analytics job you want information about. This value is case sensitive.

newGetCallAnalyticsJob

Arguments

:: Text
$sel:callAnalyticsJobName:GetCallAnalyticsJob'

-> GetCallAnalyticsJob

Create a value of GetCallAnalyticsJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:callAnalyticsJobName:GetCallAnalyticsJob', getCallAnalyticsJob_callAnalyticsJobName - The name of the analytics job you want information about. This value is case sensitive.

Request Lenses

getCallAnalyticsJob_callAnalyticsJobName :: Lens' GetCallAnalyticsJob Text

The name of the analytics job you want information about. This value is case sensitive.

Destructuring the Response

data GetCallAnalyticsJobResponse

See: newGetCallAnalyticsJobResponse smart constructor.

Constructors

GetCallAnalyticsJobResponse'

Fields

callAnalyticsJob :: Maybe CallAnalyticsJob

An object that contains the results of your call analytics job.

httpStatus :: Int

The response's http status code.

newGetCallAnalyticsJobResponse

Arguments

:: Int
$sel:httpStatus:GetCallAnalyticsJobResponse'

-> GetCallAnalyticsJobResponse

Create a value of GetCallAnalyticsJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:callAnalyticsJob:GetCallAnalyticsJobResponse', getCallAnalyticsJobResponse_callAnalyticsJob - An object that contains the results of your call analytics job.

$sel:httpStatus:GetCallAnalyticsJobResponse', getCallAnalyticsJobResponse_httpStatus - The response's http status code.

Response Lenses

getCallAnalyticsJobResponse_callAnalyticsJob :: Lens' GetCallAnalyticsJobResponse (Maybe CallAnalyticsJob)

An object that contains the results of your call analytics job.

getCallAnalyticsJobResponse_httpStatus :: Lens' GetCallAnalyticsJobResponse Int

The response's http status code.