Network.AWS.Transcribe.GetMedicalVocabulary

Description

Retrieves information about a medical vocabulary.

Creating a Request

data GetMedicalVocabulary

See: newGetMedicalVocabulary smart constructor.

Constructors

GetMedicalVocabulary'

Fields

vocabularyName :: Text

The name of the vocabulary that you want information about. The value is case sensitive.

newGetMedicalVocabulary

Arguments

:: Text
$sel:vocabularyName:GetMedicalVocabulary'

-> GetMedicalVocabulary

Create a value of GetMedicalVocabulary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vocabularyName:GetMedicalVocabulary', getMedicalVocabulary_vocabularyName - The name of the vocabulary that you want information about. The value is case sensitive.

Request Lenses

getMedicalVocabulary_vocabularyName :: Lens' GetMedicalVocabulary Text

The name of the vocabulary that you want information about. The value is case sensitive.

Destructuring the Response

data GetMedicalVocabularyResponse

See: newGetMedicalVocabularyResponse smart constructor.

Constructors

GetMedicalVocabularyResponse'

Fields

languageCode :: Maybe LanguageCode

The valid language code for your vocabulary entries.

failureReason :: Maybe Text

If the VocabularyState is FAILED, this field contains information about why the job failed.

lastModifiedTime :: Maybe POSIX

The date and time that the vocabulary was last modified with a text file different from the one that was previously used.

vocabularyName :: Maybe Text

The name of the vocabulary returned by Amazon Transcribe Medical.

vocabularyState :: Maybe VocabularyState

The processing state of the vocabulary. If the VocabularyState is READY then you can use it in the StartMedicalTranscriptionJob operation.

downloadUri :: Maybe Text

The location in Amazon S3 where the vocabulary is stored. Use this URI to get the contents of the vocabulary. You can download your vocabulary from the URI for a limited time.

httpStatus :: Int

The response's http status code.

newGetMedicalVocabularyResponse

Arguments

:: Int
$sel:httpStatus:GetMedicalVocabularyResponse'

-> GetMedicalVocabularyResponse

Create a value of GetMedicalVocabularyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:GetMedicalVocabularyResponse', getMedicalVocabularyResponse_languageCode - The valid language code for your vocabulary entries.

$sel:failureReason:GetMedicalVocabularyResponse', getMedicalVocabularyResponse_failureReason - If the VocabularyState is FAILED, this field contains information about why the job failed.

$sel:lastModifiedTime:GetMedicalVocabularyResponse', getMedicalVocabularyResponse_lastModifiedTime - The date and time that the vocabulary was last modified with a text file different from the one that was previously used.

$sel:vocabularyName:GetMedicalVocabulary', getMedicalVocabularyResponse_vocabularyName - The name of the vocabulary returned by Amazon Transcribe Medical.

$sel:vocabularyState:GetMedicalVocabularyResponse', getMedicalVocabularyResponse_vocabularyState - The processing state of the vocabulary. If the VocabularyState is READY then you can use it in the StartMedicalTranscriptionJob operation.

$sel:downloadUri:GetMedicalVocabularyResponse', getMedicalVocabularyResponse_downloadUri - The location in Amazon S3 where the vocabulary is stored. Use this URI to get the contents of the vocabulary. You can download your vocabulary from the URI for a limited time.

$sel:httpStatus:GetMedicalVocabularyResponse', getMedicalVocabularyResponse_httpStatus - The response's http status code.

Response Lenses

getMedicalVocabularyResponse_languageCode :: Lens' GetMedicalVocabularyResponse (Maybe LanguageCode)

The valid language code for your vocabulary entries.

getMedicalVocabularyResponse_failureReason :: Lens' GetMedicalVocabularyResponse (Maybe Text)

If the VocabularyState is FAILED, this field contains information about why the job failed.

getMedicalVocabularyResponse_lastModifiedTime :: Lens' GetMedicalVocabularyResponse (Maybe UTCTime)

The date and time that the vocabulary was last modified with a text file different from the one that was previously used.

getMedicalVocabularyResponse_vocabularyName :: Lens' GetMedicalVocabularyResponse (Maybe Text)

The name of the vocabulary returned by Amazon Transcribe Medical.

getMedicalVocabularyResponse_vocabularyState :: Lens' GetMedicalVocabularyResponse (Maybe VocabularyState)

The processing state of the vocabulary. If the VocabularyState is READY then you can use it in the StartMedicalTranscriptionJob operation.

getMedicalVocabularyResponse_downloadUri :: Lens' GetMedicalVocabularyResponse (Maybe Text)

The location in Amazon S3 where the vocabulary is stored. Use this URI to get the contents of the vocabulary. You can download your vocabulary from the URI for a limited time.

getMedicalVocabularyResponse_httpStatus :: Lens' GetMedicalVocabularyResponse Int

The response's http status code.