Network.AWS.Transcribe.GetVocabularyFilter

Description

Returns information about a vocabulary filter.

Creating a Request

data GetVocabularyFilter

See: newGetVocabularyFilter smart constructor.

Constructors

GetVocabularyFilter'

Fields

vocabularyFilterName :: Text

The name of the vocabulary filter for which to return information.

newGetVocabularyFilter

Arguments

:: Text
$sel:vocabularyFilterName:GetVocabularyFilter'

-> GetVocabularyFilter

Create a value of GetVocabularyFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vocabularyFilterName:GetVocabularyFilter', getVocabularyFilter_vocabularyFilterName - The name of the vocabulary filter for which to return information.

Request Lenses

getVocabularyFilter_vocabularyFilterName :: Lens' GetVocabularyFilter Text

The name of the vocabulary filter for which to return information.

Destructuring the Response

data GetVocabularyFilterResponse

See: newGetVocabularyFilterResponse smart constructor.

Constructors

GetVocabularyFilterResponse'

Fields

languageCode :: Maybe LanguageCode

The language code of the words in the vocabulary filter.

downloadUri :: Maybe Text

The URI of the list of words in the vocabulary filter. You can use this URI to get the list of words.

lastModifiedTime :: Maybe POSIX

The date and time that the contents of the vocabulary filter were updated.

vocabularyFilterName :: Maybe Text

The name of the vocabulary filter.

httpStatus :: Int

The response's http status code.

newGetVocabularyFilterResponse

Arguments

:: Int
$sel:httpStatus:GetVocabularyFilterResponse'

-> GetVocabularyFilterResponse

Create a value of GetVocabularyFilterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:GetVocabularyFilterResponse', getVocabularyFilterResponse_languageCode - The language code of the words in the vocabulary filter.

$sel:downloadUri:GetVocabularyFilterResponse', getVocabularyFilterResponse_downloadUri - The URI of the list of words in the vocabulary filter. You can use this URI to get the list of words.

$sel:lastModifiedTime:GetVocabularyFilterResponse', getVocabularyFilterResponse_lastModifiedTime - The date and time that the contents of the vocabulary filter were updated.

$sel:vocabularyFilterName:GetVocabularyFilter', getVocabularyFilterResponse_vocabularyFilterName - The name of the vocabulary filter.

$sel:httpStatus:GetVocabularyFilterResponse', getVocabularyFilterResponse_httpStatus - The response's http status code.

Response Lenses

getVocabularyFilterResponse_languageCode :: Lens' GetVocabularyFilterResponse (Maybe LanguageCode)

The language code of the words in the vocabulary filter.

getVocabularyFilterResponse_downloadUri :: Lens' GetVocabularyFilterResponse (Maybe Text)

The URI of the list of words in the vocabulary filter. You can use this URI to get the list of words.

getVocabularyFilterResponse_lastModifiedTime :: Lens' GetVocabularyFilterResponse (Maybe UTCTime)

The date and time that the contents of the vocabulary filter were updated.

getVocabularyFilterResponse_vocabularyFilterName :: Lens' GetVocabularyFilterResponse (Maybe Text)

The name of the vocabulary filter.

getVocabularyFilterResponse_httpStatus :: Lens' GetVocabularyFilterResponse Int

The response's http status code.