Network.AWS.Transcribe.ListMedicalVocabularies

Description

Returns a list of vocabularies that match the specified criteria. If you don't enter a value in any of the request parameters, returns the entire list of vocabularies.

Creating a Request

data ListMedicalVocabularies

See: newListMedicalVocabularies smart constructor.

Constructors

ListMedicalVocabularies'

Fields

nextToken :: Maybe Text

If the result of your previous request to ListMedicalVocabularies was truncated, include the NextToken to fetch the next set of vocabularies.

nameContains :: Maybe Text

Returns vocabularies whose names contain the specified string. The search is not case sensitive. ListMedicalVocabularies returns both "vocabularyname" and "VocabularyName".

maxResults :: Maybe Natural

The maximum number of vocabularies to return in each page of results. If there are fewer results than the value you specify, only the actual results are returned. If you do not specify a value, the default of 5 is used.

stateEquals :: Maybe VocabularyState

When specified, returns only vocabularies with the VocabularyState equal to the specified vocabulary state. Use this field to see which vocabularies are ready for your medical transcription jobs.

newListMedicalVocabularies :: ListMedicalVocabularies

Create a value of ListMedicalVocabularies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMedicalVocabularies', listMedicalVocabularies_nextToken - If the result of your previous request to ListMedicalVocabularies was truncated, include the NextToken to fetch the next set of vocabularies.

$sel:nameContains:ListMedicalVocabularies', listMedicalVocabularies_nameContains - Returns vocabularies whose names contain the specified string. The search is not case sensitive. ListMedicalVocabularies returns both "vocabularyname" and "VocabularyName".

$sel:maxResults:ListMedicalVocabularies', listMedicalVocabularies_maxResults - The maximum number of vocabularies to return in each page of results. If there are fewer results than the value you specify, only the actual results are returned. If you do not specify a value, the default of 5 is used.

$sel:stateEquals:ListMedicalVocabularies', listMedicalVocabularies_stateEquals - When specified, returns only vocabularies with the VocabularyState equal to the specified vocabulary state. Use this field to see which vocabularies are ready for your medical transcription jobs.

Request Lenses

listMedicalVocabularies_nextToken :: Lens' ListMedicalVocabularies (Maybe Text)

If the result of your previous request to ListMedicalVocabularies was truncated, include the NextToken to fetch the next set of vocabularies.

listMedicalVocabularies_nameContains :: Lens' ListMedicalVocabularies (Maybe Text)

Returns vocabularies whose names contain the specified string. The search is not case sensitive. ListMedicalVocabularies returns both "vocabularyname" and "VocabularyName".

listMedicalVocabularies_maxResults :: Lens' ListMedicalVocabularies (Maybe Natural)

The maximum number of vocabularies to return in each page of results. If there are fewer results than the value you specify, only the actual results are returned. If you do not specify a value, the default of 5 is used.

listMedicalVocabularies_stateEquals :: Lens' ListMedicalVocabularies (Maybe VocabularyState)

When specified, returns only vocabularies with the VocabularyState equal to the specified vocabulary state. Use this field to see which vocabularies are ready for your medical transcription jobs.

Destructuring the Response

data ListMedicalVocabulariesResponse

See: newListMedicalVocabulariesResponse smart constructor.

Constructors

ListMedicalVocabulariesResponse'

Fields

status :: Maybe VocabularyState

The requested vocabulary state.

nextToken :: Maybe Text

The ListMedicalVocabularies operation returns a page of vocabularies at a time. You set the maximum number of vocabularies to return on a page with the MaxResults parameter. If there are more jobs in the list will fit on a page, Amazon Transcribe Medical returns the NextPage token. To return the next page of vocabularies, include the token in the next request to the ListMedicalVocabularies operation .

vocabularies :: Maybe [VocabularyInfo]

A list of objects that describe the vocabularies that match your search criteria.

httpStatus :: Int

The response's http status code.

newListMedicalVocabulariesResponse

Arguments

:: Int
$sel:httpStatus:ListMedicalVocabulariesResponse'

-> ListMedicalVocabulariesResponse

Create a value of ListMedicalVocabulariesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListMedicalVocabulariesResponse', listMedicalVocabulariesResponse_status - The requested vocabulary state.

$sel:nextToken:ListMedicalVocabularies', listMedicalVocabulariesResponse_nextToken - The ListMedicalVocabularies operation returns a page of vocabularies at a time. You set the maximum number of vocabularies to return on a page with the MaxResults parameter. If there are more jobs in the list will fit on a page, Amazon Transcribe Medical returns the NextPage token. To return the next page of vocabularies, include the token in the next request to the ListMedicalVocabularies operation .

$sel:vocabularies:ListMedicalVocabulariesResponse', listMedicalVocabulariesResponse_vocabularies - A list of objects that describe the vocabularies that match your search criteria.

$sel:httpStatus:ListMedicalVocabulariesResponse', listMedicalVocabulariesResponse_httpStatus - The response's http status code.

Response Lenses

listMedicalVocabulariesResponse_status :: Lens' ListMedicalVocabulariesResponse (Maybe VocabularyState)

The requested vocabulary state.

listMedicalVocabulariesResponse_nextToken :: Lens' ListMedicalVocabulariesResponse (Maybe Text)

The ListMedicalVocabularies operation returns a page of vocabularies at a time. You set the maximum number of vocabularies to return on a page with the MaxResults parameter. If there are more jobs in the list will fit on a page, Amazon Transcribe Medical returns the NextPage token. To return the next page of vocabularies, include the token in the next request to the ListMedicalVocabularies operation .

listMedicalVocabulariesResponse_vocabularies :: Lens' ListMedicalVocabulariesResponse (Maybe [VocabularyInfo])

A list of objects that describe the vocabularies that match your search criteria.

listMedicalVocabulariesResponse_httpStatus :: Lens' ListMedicalVocabulariesResponse Int

The response's http status code.