Network.AWS.Transcribe.ListVocabularies

Description

Returns a list of vocabularies that match the specified criteria. If no criteria are specified, returns the entire list of vocabularies.

Creating a Request

data ListVocabularies

See: newListVocabularies smart constructor.

Constructors

ListVocabularies'

Fields

nextToken :: Maybe Text

If the result of the previous request to ListVocabularies was truncated, include the NextToken to fetch the next set of jobs.

nameContains :: Maybe Text

When specified, the vocabularies returned in the list are limited to vocabularies whose name contains the specified string. The search is not case sensitive, ListVocabularies returns both "vocabularyname" and "VocabularyName" in the response list.

maxResults :: Maybe Natural

The maximum number of vocabularies to return in each page of results. If there are fewer results than the value you specify, only the actual results are returned. If you do not specify a value, the default of 5 is used.

stateEquals :: Maybe VocabularyState

When specified, only returns vocabularies with the VocabularyState field equal to the specified state.

newListVocabularies :: ListVocabularies

Create a value of ListVocabularies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListVocabularies', listVocabularies_nextToken - If the result of the previous request to ListVocabularies was truncated, include the NextToken to fetch the next set of jobs.

$sel:nameContains:ListVocabularies', listVocabularies_nameContains - When specified, the vocabularies returned in the list are limited to vocabularies whose name contains the specified string. The search is not case sensitive, ListVocabularies returns both "vocabularyname" and "VocabularyName" in the response list.

$sel:maxResults:ListVocabularies', listVocabularies_maxResults - The maximum number of vocabularies to return in each page of results. If there are fewer results than the value you specify, only the actual results are returned. If you do not specify a value, the default of 5 is used.

$sel:stateEquals:ListVocabularies', listVocabularies_stateEquals - When specified, only returns vocabularies with the VocabularyState field equal to the specified state.

Request Lenses

listVocabularies_nextToken :: Lens' ListVocabularies (Maybe Text)

If the result of the previous request to ListVocabularies was truncated, include the NextToken to fetch the next set of jobs.

listVocabularies_nameContains :: Lens' ListVocabularies (Maybe Text)

When specified, the vocabularies returned in the list are limited to vocabularies whose name contains the specified string. The search is not case sensitive, ListVocabularies returns both "vocabularyname" and "VocabularyName" in the response list.

listVocabularies_maxResults :: Lens' ListVocabularies (Maybe Natural)

The maximum number of vocabularies to return in each page of results. If there are fewer results than the value you specify, only the actual results are returned. If you do not specify a value, the default of 5 is used.

listVocabularies_stateEquals :: Lens' ListVocabularies (Maybe VocabularyState)

When specified, only returns vocabularies with the VocabularyState field equal to the specified state.

Destructuring the Response

data ListVocabulariesResponse

See: newListVocabulariesResponse smart constructor.

Constructors

ListVocabulariesResponse'

Fields

status :: Maybe VocabularyState

The requested vocabulary state.

nextToken :: Maybe Text

The ListVocabularies operation returns a page of vocabularies at a time. The maximum size of the page is set in the MaxResults parameter. If there are more jobs in the list than will fit on the page, Amazon Transcribe returns the NextPage token. To return in the next page of jobs, include the token in the next request to the ListVocabularies operation.

vocabularies :: Maybe [VocabularyInfo]

A list of objects that describe the vocabularies that match the search criteria in the request.

httpStatus :: Int

The response's http status code.

newListVocabulariesResponse

Arguments

:: Int
$sel:httpStatus:ListVocabulariesResponse'

-> ListVocabulariesResponse

Create a value of ListVocabulariesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListVocabulariesResponse', listVocabulariesResponse_status - The requested vocabulary state.

$sel:nextToken:ListVocabularies', listVocabulariesResponse_nextToken - The ListVocabularies operation returns a page of vocabularies at a time. The maximum size of the page is set in the MaxResults parameter. If there are more jobs in the list than will fit on the page, Amazon Transcribe returns the NextPage token. To return in the next page of jobs, include the token in the next request to the ListVocabularies operation.

$sel:vocabularies:ListVocabulariesResponse', listVocabulariesResponse_vocabularies - A list of objects that describe the vocabularies that match the search criteria in the request.

$sel:httpStatus:ListVocabulariesResponse', listVocabulariesResponse_httpStatus - The response's http status code.

Response Lenses

listVocabulariesResponse_status :: Lens' ListVocabulariesResponse (Maybe VocabularyState)

The requested vocabulary state.

listVocabulariesResponse_nextToken :: Lens' ListVocabulariesResponse (Maybe Text)

The ListVocabularies operation returns a page of vocabularies at a time. The maximum size of the page is set in the MaxResults parameter. If there are more jobs in the list than will fit on the page, Amazon Transcribe returns the NextPage token. To return in the next page of jobs, include the token in the next request to the ListVocabularies operation.

listVocabulariesResponse_vocabularies :: Lens' ListVocabulariesResponse (Maybe [VocabularyInfo])

A list of objects that describe the vocabularies that match the search criteria in the request.

listVocabulariesResponse_httpStatus :: Lens' ListVocabulariesResponse Int

The response's http status code.