Network.AWS.Transcribe.ListVocabularyFilters

Description

Gets information about vocabulary filters.

Creating a Request

data ListVocabularyFilters

See: newListVocabularyFilters smart constructor.

Constructors

ListVocabularyFilters'

Fields

nextToken :: Maybe Text

If the result of the previous request to ListVocabularyFilters was truncated, include the NextToken to fetch the next set of collections.

nameContains :: Maybe Text

Filters the response so that it only contains vocabulary filters whose name contains the specified string.

maxResults :: Maybe Natural

The maximum number of filters to return in each page of results. If there are fewer results than the value you specify, only the actual results are returned. If you do not specify a value, the default of 5 is used.

newListVocabularyFilters :: ListVocabularyFilters

Create a value of ListVocabularyFilters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListVocabularyFilters', listVocabularyFilters_nextToken - If the result of the previous request to ListVocabularyFilters was truncated, include the NextToken to fetch the next set of collections.

$sel:nameContains:ListVocabularyFilters', listVocabularyFilters_nameContains - Filters the response so that it only contains vocabulary filters whose name contains the specified string.

$sel:maxResults:ListVocabularyFilters', listVocabularyFilters_maxResults - The maximum number of filters to return in each page of results. If there are fewer results than the value you specify, only the actual results are returned. If you do not specify a value, the default of 5 is used.

Request Lenses

listVocabularyFilters_nextToken :: Lens' ListVocabularyFilters (Maybe Text)

If the result of the previous request to ListVocabularyFilters was truncated, include the NextToken to fetch the next set of collections.

listVocabularyFilters_nameContains :: Lens' ListVocabularyFilters (Maybe Text)

Filters the response so that it only contains vocabulary filters whose name contains the specified string.

listVocabularyFilters_maxResults :: Lens' ListVocabularyFilters (Maybe Natural)

The maximum number of filters to return in each page of results. If there are fewer results than the value you specify, only the actual results are returned. If you do not specify a value, the default of 5 is used.

Destructuring the Response

data ListVocabularyFiltersResponse

See: newListVocabularyFiltersResponse smart constructor.

Constructors

ListVocabularyFiltersResponse'

Fields

nextToken :: Maybe Text

The ListVocabularyFilters operation returns a page of collections at a time. The maximum size of the page is set by the MaxResults parameter. If there are more jobs in the list than the page size, Amazon Transcribe returns the NextPage token. Include the token in the next request to the ListVocabularyFilters operation to return in the next page of jobs.

vocabularyFilters :: Maybe [VocabularyFilterInfo]

The list of vocabulary filters. It contains at most MaxResults number of filters. If there are more filters, call the ListVocabularyFilters operation again with the NextToken parameter in the request set to the value of the NextToken field in the response.

httpStatus :: Int

The response's http status code.

newListVocabularyFiltersResponse

Arguments

:: Int
$sel:httpStatus:ListVocabularyFiltersResponse'

-> ListVocabularyFiltersResponse

Create a value of ListVocabularyFiltersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListVocabularyFilters', listVocabularyFiltersResponse_nextToken - The ListVocabularyFilters operation returns a page of collections at a time. The maximum size of the page is set by the MaxResults parameter. If there are more jobs in the list than the page size, Amazon Transcribe returns the NextPage token. Include the token in the next request to the ListVocabularyFilters operation to return in the next page of jobs.

$sel:vocabularyFilters:ListVocabularyFiltersResponse', listVocabularyFiltersResponse_vocabularyFilters - The list of vocabulary filters. It contains at most MaxResults number of filters. If there are more filters, call the ListVocabularyFilters operation again with the NextToken parameter in the request set to the value of the NextToken field in the response.

$sel:httpStatus:ListVocabularyFiltersResponse', listVocabularyFiltersResponse_httpStatus - The response's http status code.

Response Lenses

listVocabularyFiltersResponse_nextToken :: Lens' ListVocabularyFiltersResponse (Maybe Text)

The ListVocabularyFilters operation returns a page of collections at a time. The maximum size of the page is set by the MaxResults parameter. If there are more jobs in the list than the page size, Amazon Transcribe returns the NextPage token. Include the token in the next request to the ListVocabularyFilters operation to return in the next page of jobs.

listVocabularyFiltersResponse_vocabularyFilters :: Lens' ListVocabularyFiltersResponse (Maybe [VocabularyFilterInfo])

The list of vocabulary filters. It contains at most MaxResults number of filters. If there are more filters, call the ListVocabularyFilters operation again with the NextToken parameter in the request set to the value of the NextToken field in the response.

listVocabularyFiltersResponse_httpStatus :: Lens' ListVocabularyFiltersResponse Int

The response's http status code.