Network.AWS.Transcribe.StartMedicalTranscriptionJob

Description

Starts a batch job to transcribe medical speech to text.

Creating a Request

data StartMedicalTranscriptionJob

See: newStartMedicalTranscriptionJob smart constructor.

Constructors

StartMedicalTranscriptionJob'

Fields

mediaFormat :: Maybe MediaFormat

The audio format of the input media file.

contentIdentificationType :: Maybe MedicalContentIdentificationType

You can configure Amazon Transcribe Medical to label content in the transcription output. If you specify PHI, Amazon Transcribe Medical labels the personal health information (PHI) that it identifies in the transcription output.

outputKey :: Maybe Text

You can specify a location in an Amazon S3 bucket to store the output of your medical transcription job.

If you don't specify an output key, Amazon Transcribe Medical stores the output of your transcription job in the Amazon S3 bucket you specified. By default, the object key is "your-transcription-job-name.json".

You can use output keys to specify the Amazon S3 prefix and file name of the transcription output. For example, specifying the Amazon S3 prefix, "folder1/folder2/", as an output key would lead to the output being stored as "folder1/folder2/your-transcription-job-name.json". If you specify "my-other-job-name.json" as the output key, the object key is changed to "my-other-job-name.json". You can use an output key to change both the prefix and the file name, for example "folder/my-other-job-name.json".

If you specify an output key, you must also specify an S3 bucket in the OutputBucketName parameter.

kmsEncryptionContext :: Maybe (HashMap Text Text)

A map of plain text, non-secret key:value pairs, known as encryption context pairs, that provide an added layer of security for your data.

outputEncryptionKMSKeyId :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon Web Services Key Management Service (KMS) key used to encrypt the output of the transcription job. The user calling the StartMedicalTranscriptionJob operation must have permission to use the specified KMS key.

You use either of the following to identify a KMS key in the current account:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

KMS Key Alias: "alias/ExampleAlias"

You can use either of the following to identify a KMS key in the current account or another account:

Amazon Resource Name (ARN) of a KMS key in the current account or another account: "arn:aws:kms:region:account ID:key/1234abcd-12ab-34cd-56ef-1234567890ab"

ARN of a KMS Key Alias: "arn:aws:kms:region:account ID:alias/ExampleAlias"

If you don't specify an encryption key, the output of the medical transcription job is encrypted with the default Amazon S3 key (SSE-S3).

If you specify a KMS key to encrypt your output, you must also specify an output location in the OutputBucketName parameter.

tags :: Maybe (NonEmpty Tag)

Add tags to an Amazon Transcribe medical transcription job.

mediaSampleRateHertz :: Maybe Natural

The sample rate, in Hertz, of the audio track in the input media file.

If you do not specify the media sample rate, Amazon Transcribe Medical determines the sample rate. If you specify the sample rate, it must match the rate detected by Amazon Transcribe Medical. In most cases, you should leave the MediaSampleRateHertz field blank and let Amazon Transcribe Medical determine the sample rate.

settings :: Maybe MedicalTranscriptionSetting

Optional settings for the medical transcription job.

medicalTranscriptionJobName :: Text

The name of the medical transcription job. You can't use the strings "." or ".." by themselves as the job name. The name must also be unique within an Amazon Web Services account. If you try to create a medical transcription job with the same name as a previous medical transcription job, you get a ConflictException error.

languageCode :: LanguageCode

The language code for the language spoken in the input media file. US English (en-US) is the valid value for medical transcription jobs. Any other value you enter for language code results in a BadRequestException error.

media :: Media

outputBucketName :: Text

The Amazon S3 location where the transcription is stored.

You must set OutputBucketName for Amazon Transcribe Medical to store the transcription results. Your transcript appears in the S3 location you specify. When you call the GetMedicalTranscriptionJob, the operation returns this location in the TranscriptFileUri field. The S3 bucket must have permissions that allow Amazon Transcribe Medical to put files in the bucket. For more information, see Permissions Required for IAM User Roles.

You can specify an Amazon Web Services Key Management Service (KMS) key to encrypt the output of your transcription using the OutputEncryptionKMSKeyId parameter. If you don't specify a KMS key, Amazon Transcribe Medical uses the default Amazon S3 key for server-side encryption of transcripts that are placed in your S3 bucket.

specialty :: Specialty

The medical specialty of any clinician speaking in the input media.

type' :: Type

The type of speech in the input audio. CONVERSATION refers to conversations between two or more speakers, e.g., a conversations between doctors and patients. DICTATION refers to single-speaker dictated speech, such as clinical notes.

newStartMedicalTranscriptionJob

Arguments

:: Text
$sel:medicalTranscriptionJobName:StartMedicalTranscriptionJob'

-> LanguageCode
$sel:languageCode:StartMedicalTranscriptionJob'

-> Media
$sel:media:StartMedicalTranscriptionJob'

-> Text
$sel:outputBucketName:StartMedicalTranscriptionJob'

-> Specialty
$sel:specialty:StartMedicalTranscriptionJob'

-> Type
$sel:type':StartMedicalTranscriptionJob'

-> StartMedicalTranscriptionJob

Create a value of StartMedicalTranscriptionJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mediaFormat:StartMedicalTranscriptionJob', startMedicalTranscriptionJob_mediaFormat - The audio format of the input media file.

$sel:contentIdentificationType:StartMedicalTranscriptionJob', startMedicalTranscriptionJob_contentIdentificationType - You can configure Amazon Transcribe Medical to label content in the transcription output. If you specify PHI, Amazon Transcribe Medical labels the personal health information (PHI) that it identifies in the transcription output.

$sel:outputKey:StartMedicalTranscriptionJob', startMedicalTranscriptionJob_outputKey - You can specify a location in an Amazon S3 bucket to store the output of your medical transcription job.

If you don't specify an output key, Amazon Transcribe Medical stores the output of your transcription job in the Amazon S3 bucket you specified. By default, the object key is "your-transcription-job-name.json".

You can use output keys to specify the Amazon S3 prefix and file name of the transcription output. For example, specifying the Amazon S3 prefix, "folder1/folder2/", as an output key would lead to the output being stored as "folder1/folder2/your-transcription-job-name.json". If you specify "my-other-job-name.json" as the output key, the object key is changed to "my-other-job-name.json". You can use an output key to change both the prefix and the file name, for example "folder/my-other-job-name.json".

If you specify an output key, you must also specify an S3 bucket in the OutputBucketName parameter.

$sel:kmsEncryptionContext:StartMedicalTranscriptionJob', startMedicalTranscriptionJob_kmsEncryptionContext - A map of plain text, non-secret key:value pairs, known as encryption context pairs, that provide an added layer of security for your data.

$sel:outputEncryptionKMSKeyId:StartMedicalTranscriptionJob', startMedicalTranscriptionJob_outputEncryptionKMSKeyId - The Amazon Resource Name (ARN) of the Amazon Web Services Key Management Service (KMS) key used to encrypt the output of the transcription job. The user calling the StartMedicalTranscriptionJob operation must have permission to use the specified KMS key.

You use either of the following to identify a KMS key in the current account:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

KMS Key Alias: "alias/ExampleAlias"

You can use either of the following to identify a KMS key in the current account or another account:

Amazon Resource Name (ARN) of a KMS key in the current account or another account: "arn:aws:kms:region:account ID:key/1234abcd-12ab-34cd-56ef-1234567890ab"

ARN of a KMS Key Alias: "arn:aws:kms:region:account ID:alias/ExampleAlias"

If you don't specify an encryption key, the output of the medical transcription job is encrypted with the default Amazon S3 key (SSE-S3).

If you specify a KMS key to encrypt your output, you must also specify an output location in the OutputBucketName parameter.

$sel:tags:StartMedicalTranscriptionJob', startMedicalTranscriptionJob_tags - Add tags to an Amazon Transcribe medical transcription job.

$sel:mediaSampleRateHertz:StartMedicalTranscriptionJob', startMedicalTranscriptionJob_mediaSampleRateHertz - The sample rate, in Hertz, of the audio track in the input media file.

If you do not specify the media sample rate, Amazon Transcribe Medical determines the sample rate. If you specify the sample rate, it must match the rate detected by Amazon Transcribe Medical. In most cases, you should leave the MediaSampleRateHertz field blank and let Amazon Transcribe Medical determine the sample rate.

$sel:settings:StartMedicalTranscriptionJob', startMedicalTranscriptionJob_settings - Optional settings for the medical transcription job.

$sel:medicalTranscriptionJobName:StartMedicalTranscriptionJob', startMedicalTranscriptionJob_medicalTranscriptionJobName - The name of the medical transcription job. You can't use the strings "." or ".." by themselves as the job name. The name must also be unique within an Amazon Web Services account. If you try to create a medical transcription job with the same name as a previous medical transcription job, you get a ConflictException error.

$sel:languageCode:StartMedicalTranscriptionJob', startMedicalTranscriptionJob_languageCode - The language code for the language spoken in the input media file. US English (en-US) is the valid value for medical transcription jobs. Any other value you enter for language code results in a BadRequestException error.

$sel:media:StartMedicalTranscriptionJob', startMedicalTranscriptionJob_media - Undocumented member.

$sel:outputBucketName:StartMedicalTranscriptionJob', startMedicalTranscriptionJob_outputBucketName - The Amazon S3 location where the transcription is stored.

You must set OutputBucketName for Amazon Transcribe Medical to store the transcription results. Your transcript appears in the S3 location you specify. When you call the GetMedicalTranscriptionJob, the operation returns this location in the TranscriptFileUri field. The S3 bucket must have permissions that allow Amazon Transcribe Medical to put files in the bucket. For more information, see Permissions Required for IAM User Roles.

You can specify an Amazon Web Services Key Management Service (KMS) key to encrypt the output of your transcription using the OutputEncryptionKMSKeyId parameter. If you don't specify a KMS key, Amazon Transcribe Medical uses the default Amazon S3 key for server-side encryption of transcripts that are placed in your S3 bucket.

$sel:specialty:StartMedicalTranscriptionJob', startMedicalTranscriptionJob_specialty - The medical specialty of any clinician speaking in the input media.

$sel:type':StartMedicalTranscriptionJob', startMedicalTranscriptionJob_type - The type of speech in the input audio. CONVERSATION refers to conversations between two or more speakers, e.g., a conversations between doctors and patients. DICTATION refers to single-speaker dictated speech, such as clinical notes.

Request Lenses

startMedicalTranscriptionJob_mediaFormat :: Lens' StartMedicalTranscriptionJob (Maybe MediaFormat)

The audio format of the input media file.

startMedicalTranscriptionJob_contentIdentificationType :: Lens' StartMedicalTranscriptionJob (Maybe MedicalContentIdentificationType)

You can configure Amazon Transcribe Medical to label content in the transcription output. If you specify PHI, Amazon Transcribe Medical labels the personal health information (PHI) that it identifies in the transcription output.

startMedicalTranscriptionJob_outputKey :: Lens' StartMedicalTranscriptionJob (Maybe Text)

You can specify a location in an Amazon S3 bucket to store the output of your medical transcription job.

If you don't specify an output key, Amazon Transcribe Medical stores the output of your transcription job in the Amazon S3 bucket you specified. By default, the object key is "your-transcription-job-name.json".

You can use output keys to specify the Amazon S3 prefix and file name of the transcription output. For example, specifying the Amazon S3 prefix, "folder1/folder2/", as an output key would lead to the output being stored as "folder1/folder2/your-transcription-job-name.json". If you specify "my-other-job-name.json" as the output key, the object key is changed to "my-other-job-name.json". You can use an output key to change both the prefix and the file name, for example "folder/my-other-job-name.json".

If you specify an output key, you must also specify an S3 bucket in the OutputBucketName parameter.

startMedicalTranscriptionJob_kmsEncryptionContext :: Lens' StartMedicalTranscriptionJob (Maybe (HashMap Text Text))

A map of plain text, non-secret key:value pairs, known as encryption context pairs, that provide an added layer of security for your data.

startMedicalTranscriptionJob_outputEncryptionKMSKeyId :: Lens' StartMedicalTranscriptionJob (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Web Services Key Management Service (KMS) key used to encrypt the output of the transcription job. The user calling the StartMedicalTranscriptionJob operation must have permission to use the specified KMS key.

You use either of the following to identify a KMS key in the current account:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

KMS Key Alias: "alias/ExampleAlias"

You can use either of the following to identify a KMS key in the current account or another account:

Amazon Resource Name (ARN) of a KMS key in the current account or another account: "arn:aws:kms:region:account ID:key/1234abcd-12ab-34cd-56ef-1234567890ab"

ARN of a KMS Key Alias: "arn:aws:kms:region:account ID:alias/ExampleAlias"

If you don't specify an encryption key, the output of the medical transcription job is encrypted with the default Amazon S3 key (SSE-S3).

If you specify a KMS key to encrypt your output, you must also specify an output location in the OutputBucketName parameter.

startMedicalTranscriptionJob_tags :: Lens' StartMedicalTranscriptionJob (Maybe (NonEmpty Tag))

Add tags to an Amazon Transcribe medical transcription job.

startMedicalTranscriptionJob_mediaSampleRateHertz :: Lens' StartMedicalTranscriptionJob (Maybe Natural)

The sample rate, in Hertz, of the audio track in the input media file.

If you do not specify the media sample rate, Amazon Transcribe Medical determines the sample rate. If you specify the sample rate, it must match the rate detected by Amazon Transcribe Medical. In most cases, you should leave the MediaSampleRateHertz field blank and let Amazon Transcribe Medical determine the sample rate.

startMedicalTranscriptionJob_settings :: Lens' StartMedicalTranscriptionJob (Maybe MedicalTranscriptionSetting)

Optional settings for the medical transcription job.

startMedicalTranscriptionJob_medicalTranscriptionJobName :: Lens' StartMedicalTranscriptionJob Text

The name of the medical transcription job. You can't use the strings "." or ".." by themselves as the job name. The name must also be unique within an Amazon Web Services account. If you try to create a medical transcription job with the same name as a previous medical transcription job, you get a ConflictException error.

startMedicalTranscriptionJob_languageCode :: Lens' StartMedicalTranscriptionJob LanguageCode

The language code for the language spoken in the input media file. US English (en-US) is the valid value for medical transcription jobs. Any other value you enter for language code results in a BadRequestException error.

startMedicalTranscriptionJob_media :: Lens' StartMedicalTranscriptionJob Media

Undocumented member.

startMedicalTranscriptionJob_outputBucketName :: Lens' StartMedicalTranscriptionJob Text

The Amazon S3 location where the transcription is stored.

You must set OutputBucketName for Amazon Transcribe Medical to store the transcription results. Your transcript appears in the S3 location you specify. When you call the GetMedicalTranscriptionJob, the operation returns this location in the TranscriptFileUri field. The S3 bucket must have permissions that allow Amazon Transcribe Medical to put files in the bucket. For more information, see Permissions Required for IAM User Roles.

You can specify an Amazon Web Services Key Management Service (KMS) key to encrypt the output of your transcription using the OutputEncryptionKMSKeyId parameter. If you don't specify a KMS key, Amazon Transcribe Medical uses the default Amazon S3 key for server-side encryption of transcripts that are placed in your S3 bucket.

startMedicalTranscriptionJob_specialty :: Lens' StartMedicalTranscriptionJob Specialty

The medical specialty of any clinician speaking in the input media.

startMedicalTranscriptionJob_type :: Lens' StartMedicalTranscriptionJob Type

The type of speech in the input audio. CONVERSATION refers to conversations between two or more speakers, e.g., a conversations between doctors and patients. DICTATION refers to single-speaker dictated speech, such as clinical notes.

Destructuring the Response

data StartMedicalTranscriptionJobResponse

See: newStartMedicalTranscriptionJobResponse smart constructor.

Constructors

StartMedicalTranscriptionJobResponse'

Fields

medicalTranscriptionJob :: Maybe MedicalTranscriptionJob

A batch job submitted to transcribe medical speech to text.

httpStatus :: Int

The response's http status code.

newStartMedicalTranscriptionJobResponse

Arguments

:: Int
$sel:httpStatus:StartMedicalTranscriptionJobResponse'

-> StartMedicalTranscriptionJobResponse

Create a value of StartMedicalTranscriptionJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:medicalTranscriptionJob:StartMedicalTranscriptionJobResponse', startMedicalTranscriptionJobResponse_medicalTranscriptionJob - A batch job submitted to transcribe medical speech to text.

$sel:httpStatus:StartMedicalTranscriptionJobResponse', startMedicalTranscriptionJobResponse_httpStatus - The response's http status code.

Response Lenses

startMedicalTranscriptionJobResponse_medicalTranscriptionJob :: Lens' StartMedicalTranscriptionJobResponse (Maybe MedicalTranscriptionJob)

A batch job submitted to transcribe medical speech to text.

startMedicalTranscriptionJobResponse_httpStatus :: Lens' StartMedicalTranscriptionJobResponse Int

The response's http status code.