Network.AWS.Transcribe.StartTranscriptionJob

Description

Starts an asynchronous job to transcribe speech to text.

Creating a Request

data StartTranscriptionJob

See: newStartTranscriptionJob smart constructor.

Constructors

StartTranscriptionJob'

Fields

languageCode :: Maybe LanguageCode

The language code for the language used in the input media file.

To transcribe speech in Modern Standard Arabic (ar-SA), your audio or video file must be encoded at a sample rate of 16,000 Hz or higher.

mediaFormat :: Maybe MediaFormat

The format of the input media file.

contentRedaction :: Maybe ContentRedaction

An object that contains the request parameters for content redaction.

identifyLanguage :: Maybe Bool

Set this field to true to enable automatic language identification. Automatic language identification is disabled by default. You receive a BadRequestException error if you enter a value for a LanguageCode.

outputKey :: Maybe Text

You can specify a location in an Amazon S3 bucket to store the output of your transcription job.

If you don't specify an output key, Amazon Transcribe stores the output of your transcription job in the Amazon S3 bucket you specified. By default, the object key is "your-transcription-job-name.json".

You can use output keys to specify the Amazon S3 prefix and file name of the transcription output. For example, specifying the Amazon S3 prefix, "folder1/folder2/", as an output key would lead to the output being stored as "folder1/folder2/your-transcription-job-name.json". If you specify "my-other-job-name.json" as the output key, the object key is changed to "my-other-job-name.json". You can use an output key to change both the prefix and the file name, for example "folder/my-other-job-name.json".

If you specify an output key, you must also specify an S3 bucket in the OutputBucketName parameter.

kmsEncryptionContext :: Maybe (HashMap Text Text)

A map of plain text, non-secret key:value pairs, known as encryption context pairs, that provide an added layer of security for your data.

modelSettings :: Maybe ModelSettings

Choose the custom language model you use for your transcription job in this parameter.

outputEncryptionKMSKeyId :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon Web Services Key Management Service (KMS) key used to encrypt the output of the transcription job. The user calling the StartTranscriptionJob operation must have permission to use the specified KMS key.

You can use either of the following to identify a KMS key in the current account:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

KMS Key Alias: "alias/ExampleAlias"

You can use either of the following to identify a KMS key in the current account or another account:

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:region:account ID:key/1234abcd-12ab-34cd-56ef-1234567890ab"

ARN of a KMS Key Alias: "arn:aws:kms:region:account ID:alias/ExampleAlias"

If you don't specify an encryption key, the output of the transcription job is encrypted with the default Amazon S3 key (SSE-S3).

If you specify a KMS key to encrypt your output, you must also specify an output location in the OutputBucketName parameter.

subtitles :: Maybe Subtitles

Add subtitles to your batch transcription job.

tags :: Maybe (NonEmpty Tag)

Add tags to an Amazon Transcribe transcription job.

mediaSampleRateHertz :: Maybe Natural

The sample rate, in Hertz, of the audio track in the input media file.

If you do not specify the media sample rate, Amazon Transcribe determines the sample rate. If you specify the sample rate, it must match the sample rate detected by Amazon Transcribe. In most cases, you should leave the MediaSampleRateHertz field blank and let Amazon Transcribe determine the sample rate.

outputBucketName :: Maybe Text

The location where the transcription is stored.

If you set the OutputBucketName, Amazon Transcribe puts the transcript in the specified S3 bucket. When you call the GetTranscriptionJob operation, the operation returns this location in the TranscriptFileUri field. If you enable content redaction, the redacted transcript appears in RedactedTranscriptFileUri. If you enable content redaction and choose to output an unredacted transcript, that transcript's location still appears in the TranscriptFileUri. The S3 bucket must have permissions that allow Amazon Transcribe to put files in the bucket. For more information, see Permissions Required for IAM User Roles.

You can specify an Amazon Web Services Key Management Service (KMS) key to encrypt the output of your transcription using the OutputEncryptionKMSKeyId parameter. If you don't specify a KMS key, Amazon Transcribe uses the default Amazon S3 key for server-side encryption of transcripts that are placed in your S3 bucket.

If you don't set the OutputBucketName, Amazon Transcribe generates a pre-signed URL, a shareable URL that provides secure access to your transcription, and returns it in the TranscriptFileUri field. Use this URL to download the transcription.

jobExecutionSettings :: Maybe JobExecutionSettings

Provides information about how a transcription job is executed. Use this field to indicate that the job can be queued for deferred execution if the concurrency limit is reached and there are no slots available to immediately run the job.

settings :: Maybe Settings

A Settings object that provides optional settings for a transcription job.

languageOptions :: Maybe (NonEmpty LanguageCode)

An object containing a list of languages that might be present in your collection of audio files. Automatic language identification chooses a language that best matches the source audio from that list.

To transcribe speech in Modern Standard Arabic (ar-SA), your audio or video file must be encoded at a sample rate of 16,000 Hz or higher.

transcriptionJobName :: Text

The name of the job. You can't use the strings "." or ".." by themselves as the job name. The name must also be unique within an Amazon Web Services account. If you try to create a transcription job with the same name as a previous transcription job, you get a ConflictException error.

media :: Media

An object that describes the input media for a transcription job.

newStartTranscriptionJob

Arguments

:: Text
$sel:transcriptionJobName:StartTranscriptionJob'

-> Media
$sel:media:StartTranscriptionJob'

-> StartTranscriptionJob

Create a value of StartTranscriptionJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:StartTranscriptionJob', startTranscriptionJob_languageCode - The language code for the language used in the input media file.

To transcribe speech in Modern Standard Arabic (ar-SA), your audio or video file must be encoded at a sample rate of 16,000 Hz or higher.

$sel:mediaFormat:StartTranscriptionJob', startTranscriptionJob_mediaFormat - The format of the input media file.

$sel:contentRedaction:StartTranscriptionJob', startTranscriptionJob_contentRedaction - An object that contains the request parameters for content redaction.

$sel:identifyLanguage:StartTranscriptionJob', startTranscriptionJob_identifyLanguage - Set this field to true to enable automatic language identification. Automatic language identification is disabled by default. You receive a BadRequestException error if you enter a value for a LanguageCode.

$sel:outputKey:StartTranscriptionJob', startTranscriptionJob_outputKey - You can specify a location in an Amazon S3 bucket to store the output of your transcription job.

If you don't specify an output key, Amazon Transcribe stores the output of your transcription job in the Amazon S3 bucket you specified. By default, the object key is "your-transcription-job-name.json".

You can use output keys to specify the Amazon S3 prefix and file name of the transcription output. For example, specifying the Amazon S3 prefix, "folder1/folder2/", as an output key would lead to the output being stored as "folder1/folder2/your-transcription-job-name.json". If you specify "my-other-job-name.json" as the output key, the object key is changed to "my-other-job-name.json". You can use an output key to change both the prefix and the file name, for example "folder/my-other-job-name.json".

If you specify an output key, you must also specify an S3 bucket in the OutputBucketName parameter.

$sel:kmsEncryptionContext:StartTranscriptionJob', startTranscriptionJob_kmsEncryptionContext - A map of plain text, non-secret key:value pairs, known as encryption context pairs, that provide an added layer of security for your data.

$sel:modelSettings:StartTranscriptionJob', startTranscriptionJob_modelSettings - Choose the custom language model you use for your transcription job in this parameter.

$sel:outputEncryptionKMSKeyId:StartTranscriptionJob', startTranscriptionJob_outputEncryptionKMSKeyId - The Amazon Resource Name (ARN) of the Amazon Web Services Key Management Service (KMS) key used to encrypt the output of the transcription job. The user calling the StartTranscriptionJob operation must have permission to use the specified KMS key.

You can use either of the following to identify a KMS key in the current account:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

KMS Key Alias: "alias/ExampleAlias"

You can use either of the following to identify a KMS key in the current account or another account:

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:region:account ID:key/1234abcd-12ab-34cd-56ef-1234567890ab"

ARN of a KMS Key Alias: "arn:aws:kms:region:account ID:alias/ExampleAlias"

If you don't specify an encryption key, the output of the transcription job is encrypted with the default Amazon S3 key (SSE-S3).

If you specify a KMS key to encrypt your output, you must also specify an output location in the OutputBucketName parameter.

$sel:subtitles:StartTranscriptionJob', startTranscriptionJob_subtitles - Add subtitles to your batch transcription job.

$sel:tags:StartTranscriptionJob', startTranscriptionJob_tags - Add tags to an Amazon Transcribe transcription job.

$sel:mediaSampleRateHertz:StartTranscriptionJob', startTranscriptionJob_mediaSampleRateHertz - The sample rate, in Hertz, of the audio track in the input media file.

If you do not specify the media sample rate, Amazon Transcribe determines the sample rate. If you specify the sample rate, it must match the sample rate detected by Amazon Transcribe. In most cases, you should leave the MediaSampleRateHertz field blank and let Amazon Transcribe determine the sample rate.

$sel:outputBucketName:StartTranscriptionJob', startTranscriptionJob_outputBucketName - The location where the transcription is stored.

If you set the OutputBucketName, Amazon Transcribe puts the transcript in the specified S3 bucket. When you call the GetTranscriptionJob operation, the operation returns this location in the TranscriptFileUri field. If you enable content redaction, the redacted transcript appears in RedactedTranscriptFileUri. If you enable content redaction and choose to output an unredacted transcript, that transcript's location still appears in the TranscriptFileUri. The S3 bucket must have permissions that allow Amazon Transcribe to put files in the bucket. For more information, see Permissions Required for IAM User Roles.

You can specify an Amazon Web Services Key Management Service (KMS) key to encrypt the output of your transcription using the OutputEncryptionKMSKeyId parameter. If you don't specify a KMS key, Amazon Transcribe uses the default Amazon S3 key for server-side encryption of transcripts that are placed in your S3 bucket.

If you don't set the OutputBucketName, Amazon Transcribe generates a pre-signed URL, a shareable URL that provides secure access to your transcription, and returns it in the TranscriptFileUri field. Use this URL to download the transcription.

$sel:jobExecutionSettings:StartTranscriptionJob', startTranscriptionJob_jobExecutionSettings - Provides information about how a transcription job is executed. Use this field to indicate that the job can be queued for deferred execution if the concurrency limit is reached and there are no slots available to immediately run the job.

$sel:settings:StartTranscriptionJob', startTranscriptionJob_settings - A Settings object that provides optional settings for a transcription job.

$sel:languageOptions:StartTranscriptionJob', startTranscriptionJob_languageOptions - An object containing a list of languages that might be present in your collection of audio files. Automatic language identification chooses a language that best matches the source audio from that list.

To transcribe speech in Modern Standard Arabic (ar-SA), your audio or video file must be encoded at a sample rate of 16,000 Hz or higher.

$sel:transcriptionJobName:StartTranscriptionJob', startTranscriptionJob_transcriptionJobName - The name of the job. You can't use the strings "." or ".." by themselves as the job name. The name must also be unique within an Amazon Web Services account. If you try to create a transcription job with the same name as a previous transcription job, you get a ConflictException error.

$sel:media:StartTranscriptionJob', startTranscriptionJob_media - An object that describes the input media for a transcription job.

Request Lenses

startTranscriptionJob_languageCode :: Lens' StartTranscriptionJob (Maybe LanguageCode)

The language code for the language used in the input media file.

To transcribe speech in Modern Standard Arabic (ar-SA), your audio or video file must be encoded at a sample rate of 16,000 Hz or higher.

startTranscriptionJob_mediaFormat :: Lens' StartTranscriptionJob (Maybe MediaFormat)

The format of the input media file.

startTranscriptionJob_contentRedaction :: Lens' StartTranscriptionJob (Maybe ContentRedaction)

An object that contains the request parameters for content redaction.

startTranscriptionJob_identifyLanguage :: Lens' StartTranscriptionJob (Maybe Bool)

Set this field to true to enable automatic language identification. Automatic language identification is disabled by default. You receive a BadRequestException error if you enter a value for a LanguageCode.

startTranscriptionJob_outputKey :: Lens' StartTranscriptionJob (Maybe Text)

You can specify a location in an Amazon S3 bucket to store the output of your transcription job.

If you don't specify an output key, Amazon Transcribe stores the output of your transcription job in the Amazon S3 bucket you specified. By default, the object key is "your-transcription-job-name.json".

You can use output keys to specify the Amazon S3 prefix and file name of the transcription output. For example, specifying the Amazon S3 prefix, "folder1/folder2/", as an output key would lead to the output being stored as "folder1/folder2/your-transcription-job-name.json". If you specify "my-other-job-name.json" as the output key, the object key is changed to "my-other-job-name.json". You can use an output key to change both the prefix and the file name, for example "folder/my-other-job-name.json".

If you specify an output key, you must also specify an S3 bucket in the OutputBucketName parameter.

startTranscriptionJob_kmsEncryptionContext :: Lens' StartTranscriptionJob (Maybe (HashMap Text Text))

A map of plain text, non-secret key:value pairs, known as encryption context pairs, that provide an added layer of security for your data.

startTranscriptionJob_modelSettings :: Lens' StartTranscriptionJob (Maybe ModelSettings)

Choose the custom language model you use for your transcription job in this parameter.

startTranscriptionJob_outputEncryptionKMSKeyId :: Lens' StartTranscriptionJob (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Web Services Key Management Service (KMS) key used to encrypt the output of the transcription job. The user calling the StartTranscriptionJob operation must have permission to use the specified KMS key.

You can use either of the following to identify a KMS key in the current account:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

KMS Key Alias: "alias/ExampleAlias"

You can use either of the following to identify a KMS key in the current account or another account:

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:region:account ID:key/1234abcd-12ab-34cd-56ef-1234567890ab"

ARN of a KMS Key Alias: "arn:aws:kms:region:account ID:alias/ExampleAlias"

If you don't specify an encryption key, the output of the transcription job is encrypted with the default Amazon S3 key (SSE-S3).

If you specify a KMS key to encrypt your output, you must also specify an output location in the OutputBucketName parameter.

startTranscriptionJob_subtitles :: Lens' StartTranscriptionJob (Maybe Subtitles)

Add subtitles to your batch transcription job.

startTranscriptionJob_tags :: Lens' StartTranscriptionJob (Maybe (NonEmpty Tag))

Add tags to an Amazon Transcribe transcription job.

startTranscriptionJob_mediaSampleRateHertz :: Lens' StartTranscriptionJob (Maybe Natural)

The sample rate, in Hertz, of the audio track in the input media file.

If you do not specify the media sample rate, Amazon Transcribe determines the sample rate. If you specify the sample rate, it must match the sample rate detected by Amazon Transcribe. In most cases, you should leave the MediaSampleRateHertz field blank and let Amazon Transcribe determine the sample rate.

startTranscriptionJob_outputBucketName :: Lens' StartTranscriptionJob (Maybe Text)

The location where the transcription is stored.

If you set the OutputBucketName, Amazon Transcribe puts the transcript in the specified S3 bucket. When you call the GetTranscriptionJob operation, the operation returns this location in the TranscriptFileUri field. If you enable content redaction, the redacted transcript appears in RedactedTranscriptFileUri. If you enable content redaction and choose to output an unredacted transcript, that transcript's location still appears in the TranscriptFileUri. The S3 bucket must have permissions that allow Amazon Transcribe to put files in the bucket. For more information, see Permissions Required for IAM User Roles.

You can specify an Amazon Web Services Key Management Service (KMS) key to encrypt the output of your transcription using the OutputEncryptionKMSKeyId parameter. If you don't specify a KMS key, Amazon Transcribe uses the default Amazon S3 key for server-side encryption of transcripts that are placed in your S3 bucket.

If you don't set the OutputBucketName, Amazon Transcribe generates a pre-signed URL, a shareable URL that provides secure access to your transcription, and returns it in the TranscriptFileUri field. Use this URL to download the transcription.

startTranscriptionJob_jobExecutionSettings :: Lens' StartTranscriptionJob (Maybe JobExecutionSettings)

Provides information about how a transcription job is executed. Use this field to indicate that the job can be queued for deferred execution if the concurrency limit is reached and there are no slots available to immediately run the job.

startTranscriptionJob_settings :: Lens' StartTranscriptionJob (Maybe Settings)

A Settings object that provides optional settings for a transcription job.

startTranscriptionJob_languageOptions :: Lens' StartTranscriptionJob (Maybe (NonEmpty LanguageCode))

An object containing a list of languages that might be present in your collection of audio files. Automatic language identification chooses a language that best matches the source audio from that list.

To transcribe speech in Modern Standard Arabic (ar-SA), your audio or video file must be encoded at a sample rate of 16,000 Hz or higher.

startTranscriptionJob_transcriptionJobName :: Lens' StartTranscriptionJob Text

The name of the job. You can't use the strings "." or ".." by themselves as the job name. The name must also be unique within an Amazon Web Services account. If you try to create a transcription job with the same name as a previous transcription job, you get a ConflictException error.

startTranscriptionJob_media :: Lens' StartTranscriptionJob Media

An object that describes the input media for a transcription job.

Destructuring the Response

data StartTranscriptionJobResponse

See: newStartTranscriptionJobResponse smart constructor.

Constructors

StartTranscriptionJobResponse'

Fields

transcriptionJob :: Maybe TranscriptionJob

An object containing details of the asynchronous transcription job.

httpStatus :: Int

The response's http status code.

newStartTranscriptionJobResponse

Arguments

:: Int
$sel:httpStatus:StartTranscriptionJobResponse'

-> StartTranscriptionJobResponse

Create a value of StartTranscriptionJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transcriptionJob:StartTranscriptionJobResponse', startTranscriptionJobResponse_transcriptionJob - An object containing details of the asynchronous transcription job.

$sel:httpStatus:StartTranscriptionJobResponse', startTranscriptionJobResponse_httpStatus - The response's http status code.

Response Lenses

startTranscriptionJobResponse_transcriptionJob :: Lens' StartTranscriptionJobResponse (Maybe TranscriptionJob)

An object containing details of the asynchronous transcription job.

startTranscriptionJobResponse_httpStatus :: Lens' StartTranscriptionJobResponse Int

The response's http status code.