Network.AWS.Transcribe.Types.CallAnalyticsJob

Description

Documentation

data CallAnalyticsJob

Describes an asynchronous analytics job that was created with the StartAnalyticsJob operation.

See: newCallAnalyticsJob smart constructor.

Constructors

CallAnalyticsJob'

Fields

creationTime :: Maybe POSIX

A timestamp that shows when the analytics job was created.

failureReason :: Maybe Text

If the AnalyticsJobStatus is FAILED, this field contains information about why the job failed.

The FailureReason field can contain one of the following values:

Unsupported media format: The media format specified in the MediaFormat field of the request isn't valid. See the description of the MediaFormat field for a list of valid values.

The media format provided does not match the detected media format: The media format of the audio file doesn't match the format specified in the MediaFormat field in the request. Check the media format of your media file and make sure the two values match.

Invalid sample rate for audio file: The sample rate specified in the MediaSampleRateHertz of the request isn't valid. The sample rate must be between 8,000 and 48,000 Hertz.

The sample rate provided does not match the detected sample rate: The sample rate in the audio file doesn't match the sample rate specified in the MediaSampleRateHertz field in the request. Check the sample rate of your media file and make sure that the two values match.

Invalid file size: file size too large: The size of your audio file is larger than what Amazon Transcribe Medical can process. For more information, see Guidelines and Quotas in the Amazon Transcribe Medical Guide.

Invalid number of channels: number of channels too large: Your audio contains more channels than Amazon Transcribe Medical is configured to process. To request additional channels, see Amazon Transcribe Medical Endpoints and Quotas in the Amazon Web Services General Reference.

callAnalyticsJobStatus :: Maybe CallAnalyticsJobStatus

The status of the analytics job.

identifiedLanguageScore :: Maybe Double

A value between zero and one that Amazon Transcribe assigned to the language that it identified in the source audio. This value appears only when you don't provide a single language code. Larger values indicate that Amazon Transcribe has higher confidence in the language that it identified

languageCode :: Maybe LanguageCode

If you know the language spoken between the customer and the agent, specify a language code for this field.

If you don't know the language, you can leave this field blank, and Amazon Transcribe will use machine learning to automatically identify the language. To improve the accuracy of language identification, you can provide an array containing the possible language codes for the language spoken in your audio. Refer to Supported languages and language-specific features for additional information.

settings :: Maybe CallAnalyticsJobSettings

Provides information about the settings used to run a transcription job.

startTime :: Maybe POSIX

A timestamp that shows when the analytics job started processing.

completionTime :: Maybe POSIX

A timestamp that shows when the analytics job was completed.

callAnalyticsJobName :: Maybe Text

The name of the call analytics job.

media :: Maybe Media

mediaFormat :: Maybe MediaFormat

The format of the input audio file. Note: for call analytics jobs, only the following media formats are supported: MP3, MP4, WAV, FLAC, OGG, and WebM.

channelDefinitions :: Maybe (NonEmpty ChannelDefinition)

Shows numeric values to indicate the channel assigned to the agent's audio and the channel assigned to the customer's audio.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Number (ARN) that you use to get access to the analytics job.

transcript :: Maybe Transcript

mediaSampleRateHertz :: Maybe Natural

The sample rate, in Hertz, of the audio.

newCallAnalyticsJob :: CallAnalyticsJob

Create a value of CallAnalyticsJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:CallAnalyticsJob', callAnalyticsJob_creationTime - A timestamp that shows when the analytics job was created.

$sel:failureReason:CallAnalyticsJob', callAnalyticsJob_failureReason - If the AnalyticsJobStatus is FAILED, this field contains information about why the job failed.

The FailureReason field can contain one of the following values:

Unsupported media format: The media format specified in the MediaFormat field of the request isn't valid. See the description of the MediaFormat field for a list of valid values.

The media format provided does not match the detected media format: The media format of the audio file doesn't match the format specified in the MediaFormat field in the request. Check the media format of your media file and make sure the two values match.

Invalid sample rate for audio file: The sample rate specified in the MediaSampleRateHertz of the request isn't valid. The sample rate must be between 8,000 and 48,000 Hertz.

The sample rate provided does not match the detected sample rate: The sample rate in the audio file doesn't match the sample rate specified in the MediaSampleRateHertz field in the request. Check the sample rate of your media file and make sure that the two values match.

Invalid file size: file size too large: The size of your audio file is larger than what Amazon Transcribe Medical can process. For more information, see Guidelines and Quotas in the Amazon Transcribe Medical Guide.

Invalid number of channels: number of channels too large: Your audio contains more channels than Amazon Transcribe Medical is configured to process. To request additional channels, see Amazon Transcribe Medical Endpoints and Quotas in the Amazon Web Services General Reference.

$sel:callAnalyticsJobStatus:CallAnalyticsJob', callAnalyticsJob_callAnalyticsJobStatus - The status of the analytics job.

$sel:identifiedLanguageScore:CallAnalyticsJob', callAnalyticsJob_identifiedLanguageScore - A value between zero and one that Amazon Transcribe assigned to the language that it identified in the source audio. This value appears only when you don't provide a single language code. Larger values indicate that Amazon Transcribe has higher confidence in the language that it identified

$sel:languageCode:CallAnalyticsJob', callAnalyticsJob_languageCode - If you know the language spoken between the customer and the agent, specify a language code for this field.

If you don't know the language, you can leave this field blank, and Amazon Transcribe will use machine learning to automatically identify the language. To improve the accuracy of language identification, you can provide an array containing the possible language codes for the language spoken in your audio. Refer to Supported languages and language-specific features for additional information.

$sel:settings:CallAnalyticsJob', callAnalyticsJob_settings - Provides information about the settings used to run a transcription job.

$sel:startTime:CallAnalyticsJob', callAnalyticsJob_startTime - A timestamp that shows when the analytics job started processing.

$sel:completionTime:CallAnalyticsJob', callAnalyticsJob_completionTime - A timestamp that shows when the analytics job was completed.

$sel:callAnalyticsJobName:CallAnalyticsJob', callAnalyticsJob_callAnalyticsJobName - The name of the call analytics job.

$sel:media:CallAnalyticsJob', callAnalyticsJob_media - Undocumented member.

$sel:mediaFormat:CallAnalyticsJob', callAnalyticsJob_mediaFormat - The format of the input audio file. Note: for call analytics jobs, only the following media formats are supported: MP3, MP4, WAV, FLAC, OGG, and WebM.

$sel:channelDefinitions:CallAnalyticsJob', callAnalyticsJob_channelDefinitions - Shows numeric values to indicate the channel assigned to the agent's audio and the channel assigned to the customer's audio.

$sel:dataAccessRoleArn:CallAnalyticsJob', callAnalyticsJob_dataAccessRoleArn - The Amazon Resource Number (ARN) that you use to get access to the analytics job.

$sel:transcript:CallAnalyticsJob', callAnalyticsJob_transcript - Undocumented member.

$sel:mediaSampleRateHertz:CallAnalyticsJob', callAnalyticsJob_mediaSampleRateHertz - The sample rate, in Hertz, of the audio.

callAnalyticsJob_creationTime :: Lens' CallAnalyticsJob (Maybe UTCTime)

A timestamp that shows when the analytics job was created.

callAnalyticsJob_failureReason :: Lens' CallAnalyticsJob (Maybe Text)

If the AnalyticsJobStatus is FAILED, this field contains information about why the job failed.

The FailureReason field can contain one of the following values:

Unsupported media format: The media format specified in the MediaFormat field of the request isn't valid. See the description of the MediaFormat field for a list of valid values.

The media format provided does not match the detected media format: The media format of the audio file doesn't match the format specified in the MediaFormat field in the request. Check the media format of your media file and make sure the two values match.

Invalid sample rate for audio file: The sample rate specified in the MediaSampleRateHertz of the request isn't valid. The sample rate must be between 8,000 and 48,000 Hertz.

The sample rate provided does not match the detected sample rate: The sample rate in the audio file doesn't match the sample rate specified in the MediaSampleRateHertz field in the request. Check the sample rate of your media file and make sure that the two values match.

Invalid file size: file size too large: The size of your audio file is larger than what Amazon Transcribe Medical can process. For more information, see Guidelines and Quotas in the Amazon Transcribe Medical Guide.

Invalid number of channels: number of channels too large: Your audio contains more channels than Amazon Transcribe Medical is configured to process. To request additional channels, see Amazon Transcribe Medical Endpoints and Quotas in the Amazon Web Services General Reference.

callAnalyticsJob_callAnalyticsJobStatus :: Lens' CallAnalyticsJob (Maybe CallAnalyticsJobStatus)

The status of the analytics job.

callAnalyticsJob_identifiedLanguageScore :: Lens' CallAnalyticsJob (Maybe Double)

A value between zero and one that Amazon Transcribe assigned to the language that it identified in the source audio. This value appears only when you don't provide a single language code. Larger values indicate that Amazon Transcribe has higher confidence in the language that it identified

callAnalyticsJob_languageCode :: Lens' CallAnalyticsJob (Maybe LanguageCode)

If you know the language spoken between the customer and the agent, specify a language code for this field.

If you don't know the language, you can leave this field blank, and Amazon Transcribe will use machine learning to automatically identify the language. To improve the accuracy of language identification, you can provide an array containing the possible language codes for the language spoken in your audio. Refer to Supported languages and language-specific features for additional information.

callAnalyticsJob_settings :: Lens' CallAnalyticsJob (Maybe CallAnalyticsJobSettings)

Provides information about the settings used to run a transcription job.

callAnalyticsJob_startTime :: Lens' CallAnalyticsJob (Maybe UTCTime)

A timestamp that shows when the analytics job started processing.

callAnalyticsJob_completionTime :: Lens' CallAnalyticsJob (Maybe UTCTime)

A timestamp that shows when the analytics job was completed.

callAnalyticsJob_callAnalyticsJobName :: Lens' CallAnalyticsJob (Maybe Text)

The name of the call analytics job.

callAnalyticsJob_media :: Lens' CallAnalyticsJob (Maybe Media)

Undocumented member.

callAnalyticsJob_mediaFormat :: Lens' CallAnalyticsJob (Maybe MediaFormat)

The format of the input audio file. Note: for call analytics jobs, only the following media formats are supported: MP3, MP4, WAV, FLAC, OGG, and WebM.

callAnalyticsJob_channelDefinitions :: Lens' CallAnalyticsJob (Maybe (NonEmpty ChannelDefinition))

Shows numeric values to indicate the channel assigned to the agent's audio and the channel assigned to the customer's audio.

callAnalyticsJob_dataAccessRoleArn :: Lens' CallAnalyticsJob (Maybe Text)

The Amazon Resource Number (ARN) that you use to get access to the analytics job.

callAnalyticsJob_transcript :: Lens' CallAnalyticsJob (Maybe Transcript)

Undocumented member.

callAnalyticsJob_mediaSampleRateHertz :: Lens' CallAnalyticsJob (Maybe Natural)

The sample rate, in Hertz, of the audio.