Network.AWS.Transcribe.Types.CallAnalyticsJobSettings

Description

Documentation

data CallAnalyticsJobSettings

Provides optional settings for the CallAnalyticsJob operation.

See: newCallAnalyticsJobSettings smart constructor.

Constructors

CallAnalyticsJobSettings'

Fields

contentRedaction :: Maybe ContentRedaction

vocabularyFilterName :: Maybe Text

The name of the vocabulary filter to use when running a call analytics job. The filter that you specify must have the same language code as the analytics job.

vocabularyFilterMethod :: Maybe VocabularyFilterMethod

Set to mask to remove filtered text from the transcript and replace it with three asterisks ("***") as placeholder text. Set to remove to remove filtered text from the transcript without using placeholder text. Set to tag to mark the word in the transcription output that matches the vocabulary filter. When you set the filter method to tag, the words matching your vocabulary filter are not masked or removed.

languageModelName :: Maybe Text

The structure used to describe a custom language model.

vocabularyName :: Maybe Text

The name of a vocabulary to use when processing the call analytics job.

languageOptions :: Maybe (NonEmpty LanguageCode)

When you run a call analytics job, you can specify the language spoken in the audio, or you can have Amazon Transcribe identify the language for you.

To specify a language, specify an array with one language code. If you don't know the language, you can leave this field blank and Amazon Transcribe will use machine learning to identify the language for you. To improve the ability of Amazon Transcribe to correctly identify the language, you can provide an array of the languages that can be present in the audio. Refer to Supported languages and language-specific features for additional information.

newCallAnalyticsJobSettings :: CallAnalyticsJobSettings

Create a value of CallAnalyticsJobSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentRedaction:CallAnalyticsJobSettings', callAnalyticsJobSettings_contentRedaction - Undocumented member.

$sel:vocabularyFilterName:CallAnalyticsJobSettings', callAnalyticsJobSettings_vocabularyFilterName - The name of the vocabulary filter to use when running a call analytics job. The filter that you specify must have the same language code as the analytics job.

$sel:vocabularyFilterMethod:CallAnalyticsJobSettings', callAnalyticsJobSettings_vocabularyFilterMethod - Set to mask to remove filtered text from the transcript and replace it with three asterisks ("***") as placeholder text. Set to remove to remove filtered text from the transcript without using placeholder text. Set to tag to mark the word in the transcription output that matches the vocabulary filter. When you set the filter method to tag, the words matching your vocabulary filter are not masked or removed.

$sel:languageModelName:CallAnalyticsJobSettings', callAnalyticsJobSettings_languageModelName - The structure used to describe a custom language model.

$sel:vocabularyName:CallAnalyticsJobSettings', callAnalyticsJobSettings_vocabularyName - The name of a vocabulary to use when processing the call analytics job.

$sel:languageOptions:CallAnalyticsJobSettings', callAnalyticsJobSettings_languageOptions - When you run a call analytics job, you can specify the language spoken in the audio, or you can have Amazon Transcribe identify the language for you.

To specify a language, specify an array with one language code. If you don't know the language, you can leave this field blank and Amazon Transcribe will use machine learning to identify the language for you. To improve the ability of Amazon Transcribe to correctly identify the language, you can provide an array of the languages that can be present in the audio. Refer to Supported languages and language-specific features for additional information.

callAnalyticsJobSettings_contentRedaction :: Lens' CallAnalyticsJobSettings (Maybe ContentRedaction)

Undocumented member.

callAnalyticsJobSettings_vocabularyFilterName :: Lens' CallAnalyticsJobSettings (Maybe Text)

The name of the vocabulary filter to use when running a call analytics job. The filter that you specify must have the same language code as the analytics job.

callAnalyticsJobSettings_vocabularyFilterMethod :: Lens' CallAnalyticsJobSettings (Maybe VocabularyFilterMethod)

Set to mask to remove filtered text from the transcript and replace it with three asterisks ("***") as placeholder text. Set to remove to remove filtered text from the transcript without using placeholder text. Set to tag to mark the word in the transcription output that matches the vocabulary filter. When you set the filter method to tag, the words matching your vocabulary filter are not masked or removed.

callAnalyticsJobSettings_languageModelName :: Lens' CallAnalyticsJobSettings (Maybe Text)

The structure used to describe a custom language model.

callAnalyticsJobSettings_vocabularyName :: Lens' CallAnalyticsJobSettings (Maybe Text)

The name of a vocabulary to use when processing the call analytics job.

callAnalyticsJobSettings_languageOptions :: Lens' CallAnalyticsJobSettings (Maybe (NonEmpty LanguageCode))

When you run a call analytics job, you can specify the language spoken in the audio, or you can have Amazon Transcribe identify the language for you.

To specify a language, specify an array with one language code. If you don't know the language, you can leave this field blank and Amazon Transcribe will use machine learning to identify the language for you. To improve the ability of Amazon Transcribe to correctly identify the language, you can provide an array of the languages that can be present in the audio. Refer to Supported languages and language-specific features for additional information.