Network.AWS.Transcribe.Types.ContentRedaction

Description

Documentation

data ContentRedaction

Settings for content redaction within a transcription job.

See: newContentRedaction smart constructor.

Constructors

ContentRedaction'

Fields

redactionType :: RedactionType

Request parameter that defines the entities to be redacted. The only accepted value is PII.

redactionOutput :: RedactionOutput

The output transcript file stored in either the default S3 bucket or in a bucket you specify.

When you choose redacted Amazon Transcribe outputs only the redacted transcript.

When you choose redacted_and_unredacted Amazon Transcribe outputs both the redacted and unredacted transcripts.

newContentRedaction

Arguments

:: RedactionType
$sel:redactionType:ContentRedaction'

-> RedactionOutput
$sel:redactionOutput:ContentRedaction'

-> ContentRedaction

Create a value of ContentRedaction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:redactionType:ContentRedaction', contentRedaction_redactionType - Request parameter that defines the entities to be redacted. The only accepted value is PII.

$sel:redactionOutput:ContentRedaction', contentRedaction_redactionOutput - The output transcript file stored in either the default S3 bucket or in a bucket you specify.

When you choose redacted Amazon Transcribe outputs only the redacted transcript.

When you choose redacted_and_unredacted Amazon Transcribe outputs both the redacted and unredacted transcripts.

contentRedaction_redactionType :: Lens' ContentRedaction RedactionType

Request parameter that defines the entities to be redacted. The only accepted value is PII.

contentRedaction_redactionOutput :: Lens' ContentRedaction RedactionOutput

The output transcript file stored in either the default S3 bucket or in a bucket you specify.

When you choose redacted Amazon Transcribe outputs only the redacted transcript.

When you choose redacted_and_unredacted Amazon Transcribe outputs both the redacted and unredacted transcripts.