Network.AWS.Transcribe.Types.MedicalTranscriptionJob

Description

Documentation

data MedicalTranscriptionJob

The data structure that contains the information for a medical transcription job.

See: newMedicalTranscriptionJob smart constructor.

Constructors

MedicalTranscriptionJob'

Fields

creationTime :: Maybe POSIX

A timestamp that shows when the job was created.

specialty :: Maybe Specialty

The medical specialty of any clinicians providing a dictation or having a conversation. Refer to Transcribing a medical conversationfor a list of supported specialties.

failureReason :: Maybe Text

If the TranscriptionJobStatus field is FAILED, this field contains information about why the job failed.

The FailureReason field contains one of the following values:

Unsupported media format- The media format specified in the MediaFormat field of the request isn't valid. See the description of the MediaFormat field for a list of valid values.

The media format provided does not match the detected media format- The media format of the audio file doesn't match the format specified in the MediaFormat field in the request. Check the media format of your media file and make sure the two values match.

Invalid sample rate for audio file- The sample rate specified in the MediaSampleRateHertz of the request isn't valid. The sample rate must be between 8,000 and 48,000 Hertz.

The sample rate provided does not match the detected sample rate- The sample rate in the audio file doesn't match the sample rate specified in the MediaSampleRateHertz field in the request. Check the sample rate of your media file and make sure that the two values match.

Invalid file size: file size too large- The size of your audio file is larger than what Amazon Transcribe Medical can process. For more information, see Guidelines and Quotas in the Amazon Transcribe Medical Guide

Invalid number of channels: number of channels too large- Your audio contains more channels than Amazon Transcribe Medical is configured to process. To request additional channels, see Amazon Transcribe Medical Endpoints and Quotas in the Amazon Web Services General Reference

languageCode :: Maybe LanguageCode

The language code for the language spoken in the source audio file. US English (en-US) is the only supported language for medical transcriptions. Any other value you enter for language code results in a BadRequestException error.

settings :: Maybe MedicalTranscriptionSetting

Object that contains object.

startTime :: Maybe POSIX

A timestamp that shows when the job started processing.

completionTime :: Maybe POSIX

A timestamp that shows when the job was completed.

media :: Maybe Media

mediaFormat :: Maybe MediaFormat

The format of the input media file.

medicalTranscriptionJobName :: Maybe Text

The name for a given medical transcription job.

transcriptionJobStatus :: Maybe TranscriptionJobStatus

The completion status of a medical transcription job.

type' :: Maybe Type

The type of speech in the transcription job. CONVERSATION is generally used for patient-physician dialogues. DICTATION is the setting for physicians speaking their notes after seeing a patient. For more information, see What is Amazon Transcribe Medical?.

contentIdentificationType :: Maybe MedicalContentIdentificationType

Shows the type of content that you've configured Amazon Transcribe Medical to identify in a transcription job. If the value is PHI, you've configured the job to identify personal health information (PHI) in the transcription output.

transcript :: Maybe MedicalTranscript

An object that contains the MedicalTranscript. The MedicalTranscript contains the TranscriptFileUri.

tags :: Maybe (NonEmpty Tag)

A key:value pair assigned to a given medical transcription job.

mediaSampleRateHertz :: Maybe Natural

The sample rate, in Hertz, of the source audio containing medical information.

If you don't specify the sample rate, Amazon Transcribe Medical determines it for you. If you choose to specify the sample rate, it must match the rate detected by Amazon Transcribe Medical. In most cases, you should leave the MedicalMediaSampleHertz blank and let Amazon Transcribe Medical determine the sample rate.

newMedicalTranscriptionJob :: MedicalTranscriptionJob

Create a value of MedicalTranscriptionJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:MedicalTranscriptionJob', medicalTranscriptionJob_creationTime - A timestamp that shows when the job was created.

$sel:specialty:MedicalTranscriptionJob', medicalTranscriptionJob_specialty - The medical specialty of any clinicians providing a dictation or having a conversation. Refer to Transcribing a medical conversationfor a list of supported specialties.

$sel:failureReason:MedicalTranscriptionJob', medicalTranscriptionJob_failureReason - If the TranscriptionJobStatus field is FAILED, this field contains information about why the job failed.

The FailureReason field contains one of the following values:

Unsupported media format- The media format specified in the MediaFormat field of the request isn't valid. See the description of the MediaFormat field for a list of valid values.

The media format provided does not match the detected media format- The media format of the audio file doesn't match the format specified in the MediaFormat field in the request. Check the media format of your media file and make sure the two values match.

Invalid sample rate for audio file- The sample rate specified in the MediaSampleRateHertz of the request isn't valid. The sample rate must be between 8,000 and 48,000 Hertz.

The sample rate provided does not match the detected sample rate- The sample rate in the audio file doesn't match the sample rate specified in the MediaSampleRateHertz field in the request. Check the sample rate of your media file and make sure that the two values match.

Invalid file size: file size too large- The size of your audio file is larger than what Amazon Transcribe Medical can process. For more information, see Guidelines and Quotas in the Amazon Transcribe Medical Guide

Invalid number of channels: number of channels too large- Your audio contains more channels than Amazon Transcribe Medical is configured to process. To request additional channels, see Amazon Transcribe Medical Endpoints and Quotas in the Amazon Web Services General Reference

$sel:languageCode:MedicalTranscriptionJob', medicalTranscriptionJob_languageCode - The language code for the language spoken in the source audio file. US English (en-US) is the only supported language for medical transcriptions. Any other value you enter for language code results in a BadRequestException error.

$sel:settings:MedicalTranscriptionJob', medicalTranscriptionJob_settings - Object that contains object.

$sel:startTime:MedicalTranscriptionJob', medicalTranscriptionJob_startTime - A timestamp that shows when the job started processing.

$sel:completionTime:MedicalTranscriptionJob', medicalTranscriptionJob_completionTime - A timestamp that shows when the job was completed.

$sel:media:MedicalTranscriptionJob', medicalTranscriptionJob_media - Undocumented member.

$sel:mediaFormat:MedicalTranscriptionJob', medicalTranscriptionJob_mediaFormat - The format of the input media file.

$sel:medicalTranscriptionJobName:MedicalTranscriptionJob', medicalTranscriptionJob_medicalTranscriptionJobName - The name for a given medical transcription job.

$sel:transcriptionJobStatus:MedicalTranscriptionJob', medicalTranscriptionJob_transcriptionJobStatus - The completion status of a medical transcription job.

$sel:type':MedicalTranscriptionJob', medicalTranscriptionJob_type - The type of speech in the transcription job. CONVERSATION is generally used for patient-physician dialogues. DICTATION is the setting for physicians speaking their notes after seeing a patient. For more information, see What is Amazon Transcribe Medical?.

$sel:contentIdentificationType:MedicalTranscriptionJob', medicalTranscriptionJob_contentIdentificationType - Shows the type of content that you've configured Amazon Transcribe Medical to identify in a transcription job. If the value is PHI, you've configured the job to identify personal health information (PHI) in the transcription output.

$sel:transcript:MedicalTranscriptionJob', medicalTranscriptionJob_transcript - An object that contains the MedicalTranscript. The MedicalTranscript contains the TranscriptFileUri.

$sel:tags:MedicalTranscriptionJob', medicalTranscriptionJob_tags - A key:value pair assigned to a given medical transcription job.

$sel:mediaSampleRateHertz:MedicalTranscriptionJob', medicalTranscriptionJob_mediaSampleRateHertz - The sample rate, in Hertz, of the source audio containing medical information.

If you don't specify the sample rate, Amazon Transcribe Medical determines it for you. If you choose to specify the sample rate, it must match the rate detected by Amazon Transcribe Medical. In most cases, you should leave the MedicalMediaSampleHertz blank and let Amazon Transcribe Medical determine the sample rate.

medicalTranscriptionJob_creationTime :: Lens' MedicalTranscriptionJob (Maybe UTCTime)

A timestamp that shows when the job was created.

medicalTranscriptionJob_specialty :: Lens' MedicalTranscriptionJob (Maybe Specialty)

The medical specialty of any clinicians providing a dictation or having a conversation. Refer to Transcribing a medical conversationfor a list of supported specialties.

medicalTranscriptionJob_failureReason :: Lens' MedicalTranscriptionJob (Maybe Text)

If the TranscriptionJobStatus field is FAILED, this field contains information about why the job failed.

The FailureReason field contains one of the following values:

Unsupported media format- The media format specified in the MediaFormat field of the request isn't valid. See the description of the MediaFormat field for a list of valid values.

The media format provided does not match the detected media format- The media format of the audio file doesn't match the format specified in the MediaFormat field in the request. Check the media format of your media file and make sure the two values match.

Invalid sample rate for audio file- The sample rate specified in the MediaSampleRateHertz of the request isn't valid. The sample rate must be between 8,000 and 48,000 Hertz.

The sample rate provided does not match the detected sample rate- The sample rate in the audio file doesn't match the sample rate specified in the MediaSampleRateHertz field in the request. Check the sample rate of your media file and make sure that the two values match.

Invalid file size: file size too large- The size of your audio file is larger than what Amazon Transcribe Medical can process. For more information, see Guidelines and Quotas in the Amazon Transcribe Medical Guide

Invalid number of channels: number of channels too large- Your audio contains more channels than Amazon Transcribe Medical is configured to process. To request additional channels, see Amazon Transcribe Medical Endpoints and Quotas in the Amazon Web Services General Reference

medicalTranscriptionJob_languageCode :: Lens' MedicalTranscriptionJob (Maybe LanguageCode)

The language code for the language spoken in the source audio file. US English (en-US) is the only supported language for medical transcriptions. Any other value you enter for language code results in a BadRequestException error.

medicalTranscriptionJob_settings :: Lens' MedicalTranscriptionJob (Maybe MedicalTranscriptionSetting)

Object that contains object.

medicalTranscriptionJob_startTime :: Lens' MedicalTranscriptionJob (Maybe UTCTime)

A timestamp that shows when the job started processing.

medicalTranscriptionJob_completionTime :: Lens' MedicalTranscriptionJob (Maybe UTCTime)

A timestamp that shows when the job was completed.

medicalTranscriptionJob_media :: Lens' MedicalTranscriptionJob (Maybe Media)

Undocumented member.

medicalTranscriptionJob_mediaFormat :: Lens' MedicalTranscriptionJob (Maybe MediaFormat)

The format of the input media file.

medicalTranscriptionJob_medicalTranscriptionJobName :: Lens' MedicalTranscriptionJob (Maybe Text)

The name for a given medical transcription job.

medicalTranscriptionJob_transcriptionJobStatus :: Lens' MedicalTranscriptionJob (Maybe TranscriptionJobStatus)

The completion status of a medical transcription job.

medicalTranscriptionJob_type :: Lens' MedicalTranscriptionJob (Maybe Type)

The type of speech in the transcription job. CONVERSATION is generally used for patient-physician dialogues. DICTATION is the setting for physicians speaking their notes after seeing a patient. For more information, see What is Amazon Transcribe Medical?.

medicalTranscriptionJob_contentIdentificationType :: Lens' MedicalTranscriptionJob (Maybe MedicalContentIdentificationType)

Shows the type of content that you've configured Amazon Transcribe Medical to identify in a transcription job. If the value is PHI, you've configured the job to identify personal health information (PHI) in the transcription output.

medicalTranscriptionJob_transcript :: Lens' MedicalTranscriptionJob (Maybe MedicalTranscript)

An object that contains the MedicalTranscript. The MedicalTranscript contains the TranscriptFileUri.

medicalTranscriptionJob_tags :: Lens' MedicalTranscriptionJob (Maybe (NonEmpty Tag))

A key:value pair assigned to a given medical transcription job.

medicalTranscriptionJob_mediaSampleRateHertz :: Lens' MedicalTranscriptionJob (Maybe Natural)

The sample rate, in Hertz, of the source audio containing medical information.

If you don't specify the sample rate, Amazon Transcribe Medical determines it for you. If you choose to specify the sample rate, it must match the rate detected by Amazon Transcribe Medical. In most cases, you should leave the MedicalMediaSampleHertz blank and let Amazon Transcribe Medical determine the sample rate.