Network.AWS.Transcribe.Types.MedicalTranscriptionJobSummary

Description

Documentation

data MedicalTranscriptionJobSummary

Provides summary information about a transcription job.

See: newMedicalTranscriptionJobSummary smart constructor.

Constructors

MedicalTranscriptionJobSummary'

Fields

creationTime :: Maybe POSIX

A timestamp that shows when the medical transcription job was created.

specialty :: Maybe Specialty

The medical specialty of the transcription job. Refer to Transcribing a medical conversationfor a list of supported specialties.

failureReason :: Maybe Text

If the TranscriptionJobStatus field is FAILED, a description of the error.

languageCode :: Maybe LanguageCode

The language of the transcript in the source audio file.

outputLocationType :: Maybe OutputLocationType

Indicates the location of the transcription job's output. This field must be the path of an S3 bucket; if you don't already have an S3 bucket, one is created based on the path you add.

startTime :: Maybe POSIX

A timestamp that shows when the job began processing.

completionTime :: Maybe POSIX

A timestamp that shows when the job was completed.

medicalTranscriptionJobName :: Maybe Text

The name of a medical transcription job.

transcriptionJobStatus :: Maybe TranscriptionJobStatus

The status of the medical transcription job.

type' :: Maybe Type

The speech of the clinician in the input audio.

contentIdentificationType :: Maybe MedicalContentIdentificationType

Shows the type of information you've configured Amazon Transcribe Medical to identify in a transcription job. If the value is PHI, you've configured the transcription job to identify personal health information (PHI).

newMedicalTranscriptionJobSummary :: MedicalTranscriptionJobSummary

Create a value of MedicalTranscriptionJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:MedicalTranscriptionJobSummary', medicalTranscriptionJobSummary_creationTime - A timestamp that shows when the medical transcription job was created.

$sel:specialty:MedicalTranscriptionJobSummary', medicalTranscriptionJobSummary_specialty - The medical specialty of the transcription job. Refer to Transcribing a medical conversationfor a list of supported specialties.

$sel:failureReason:MedicalTranscriptionJobSummary', medicalTranscriptionJobSummary_failureReason - If the TranscriptionJobStatus field is FAILED, a description of the error.

$sel:languageCode:MedicalTranscriptionJobSummary', medicalTranscriptionJobSummary_languageCode - The language of the transcript in the source audio file.

$sel:outputLocationType:MedicalTranscriptionJobSummary', medicalTranscriptionJobSummary_outputLocationType - Indicates the location of the transcription job's output. This field must be the path of an S3 bucket; if you don't already have an S3 bucket, one is created based on the path you add.

$sel:startTime:MedicalTranscriptionJobSummary', medicalTranscriptionJobSummary_startTime - A timestamp that shows when the job began processing.

$sel:completionTime:MedicalTranscriptionJobSummary', medicalTranscriptionJobSummary_completionTime - A timestamp that shows when the job was completed.

$sel:medicalTranscriptionJobName:MedicalTranscriptionJobSummary', medicalTranscriptionJobSummary_medicalTranscriptionJobName - The name of a medical transcription job.

$sel:transcriptionJobStatus:MedicalTranscriptionJobSummary', medicalTranscriptionJobSummary_transcriptionJobStatus - The status of the medical transcription job.

$sel:type':MedicalTranscriptionJobSummary', medicalTranscriptionJobSummary_type - The speech of the clinician in the input audio.

$sel:contentIdentificationType:MedicalTranscriptionJobSummary', medicalTranscriptionJobSummary_contentIdentificationType - Shows the type of information you've configured Amazon Transcribe Medical to identify in a transcription job. If the value is PHI, you've configured the transcription job to identify personal health information (PHI).

medicalTranscriptionJobSummary_creationTime :: Lens' MedicalTranscriptionJobSummary (Maybe UTCTime)

A timestamp that shows when the medical transcription job was created.

medicalTranscriptionJobSummary_specialty :: Lens' MedicalTranscriptionJobSummary (Maybe Specialty)

The medical specialty of the transcription job. Refer to Transcribing a medical conversationfor a list of supported specialties.

medicalTranscriptionJobSummary_failureReason :: Lens' MedicalTranscriptionJobSummary (Maybe Text)

If the TranscriptionJobStatus field is FAILED, a description of the error.

medicalTranscriptionJobSummary_languageCode :: Lens' MedicalTranscriptionJobSummary (Maybe LanguageCode)

The language of the transcript in the source audio file.

medicalTranscriptionJobSummary_outputLocationType :: Lens' MedicalTranscriptionJobSummary (Maybe OutputLocationType)

Indicates the location of the transcription job's output. This field must be the path of an S3 bucket; if you don't already have an S3 bucket, one is created based on the path you add.

medicalTranscriptionJobSummary_startTime :: Lens' MedicalTranscriptionJobSummary (Maybe UTCTime)

A timestamp that shows when the job began processing.

medicalTranscriptionJobSummary_completionTime :: Lens' MedicalTranscriptionJobSummary (Maybe UTCTime)

A timestamp that shows when the job was completed.

medicalTranscriptionJobSummary_medicalTranscriptionJobName :: Lens' MedicalTranscriptionJobSummary (Maybe Text)

The name of a medical transcription job.

medicalTranscriptionJobSummary_transcriptionJobStatus :: Lens' MedicalTranscriptionJobSummary (Maybe TranscriptionJobStatus)

The status of the medical transcription job.

medicalTranscriptionJobSummary_type :: Lens' MedicalTranscriptionJobSummary (Maybe Type)

The speech of the clinician in the input audio.

medicalTranscriptionJobSummary_contentIdentificationType :: Lens' MedicalTranscriptionJobSummary (Maybe MedicalContentIdentificationType)

Shows the type of information you've configured Amazon Transcribe Medical to identify in a transcription job. If the value is PHI, you've configured the transcription job to identify personal health information (PHI).