Network.AWS.Transcribe.Types.MedicalTranscriptionSetting

Description

Documentation

data MedicalTranscriptionSetting

Optional settings for the StartMedicalTranscriptionJob operation.

See: newMedicalTranscriptionSetting smart constructor.

Constructors

MedicalTranscriptionSetting'

Fields

showAlternatives :: Maybe Bool

Determines whether alternative transcripts are generated along with the transcript that has the highest confidence. If you set ShowAlternatives field to true, you must also set the maximum number of alternatives to return in the MaxAlternatives field.

channelIdentification :: Maybe Bool

Instructs Amazon Transcribe Medical to process each audio channel separately and then merge the transcription output of each channel into a single transcription.

Amazon Transcribe Medical also produces a transcription of each item detected on an audio channel, including the start time and end time of the item and alternative transcriptions of item. The alternative transcriptions also come with confidence scores provided by Amazon Transcribe Medical.

You can't set both ShowSpeakerLabels and ChannelIdentification in the same request. If you set both, your request returns a BadRequestException

maxAlternatives :: Maybe Natural

The maximum number of alternatives that you tell the service to return. If you specify the MaxAlternatives field, you must set the ShowAlternatives field to true.

showSpeakerLabels :: Maybe Bool

Determines whether the transcription job uses speaker recognition to identify different speakers in the input audio. Speaker recognition labels individual speakers in the audio file. If you set the ShowSpeakerLabels field to true, you must also set the maximum number of speaker labels in the MaxSpeakerLabels field.

You can't set both ShowSpeakerLabels and ChannelIdentification in the same request. If you set both, your request returns a BadRequestException.

vocabularyName :: Maybe Text

The name of the vocabulary to use when processing a medical transcription job.

maxSpeakerLabels :: Maybe Natural

The maximum number of speakers to identify in the input audio. If there are more speakers in the audio than this number, multiple speakers are identified as a single speaker. If you specify the MaxSpeakerLabels field, you must set the ShowSpeakerLabels field to true.

newMedicalTranscriptionSetting :: MedicalTranscriptionSetting

Create a value of MedicalTranscriptionSetting with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:showAlternatives:MedicalTranscriptionSetting', medicalTranscriptionSetting_showAlternatives - Determines whether alternative transcripts are generated along with the transcript that has the highest confidence. If you set ShowAlternatives field to true, you must also set the maximum number of alternatives to return in the MaxAlternatives field.

$sel:channelIdentification:MedicalTranscriptionSetting', medicalTranscriptionSetting_channelIdentification - Instructs Amazon Transcribe Medical to process each audio channel separately and then merge the transcription output of each channel into a single transcription.

Amazon Transcribe Medical also produces a transcription of each item detected on an audio channel, including the start time and end time of the item and alternative transcriptions of item. The alternative transcriptions also come with confidence scores provided by Amazon Transcribe Medical.

You can't set both ShowSpeakerLabels and ChannelIdentification in the same request. If you set both, your request returns a BadRequestException

$sel:maxAlternatives:MedicalTranscriptionSetting', medicalTranscriptionSetting_maxAlternatives - The maximum number of alternatives that you tell the service to return. If you specify the MaxAlternatives field, you must set the ShowAlternatives field to true.

$sel:showSpeakerLabels:MedicalTranscriptionSetting', medicalTranscriptionSetting_showSpeakerLabels - Determines whether the transcription job uses speaker recognition to identify different speakers in the input audio. Speaker recognition labels individual speakers in the audio file. If you set the ShowSpeakerLabels field to true, you must also set the maximum number of speaker labels in the MaxSpeakerLabels field.

You can't set both ShowSpeakerLabels and ChannelIdentification in the same request. If you set both, your request returns a BadRequestException.

$sel:vocabularyName:MedicalTranscriptionSetting', medicalTranscriptionSetting_vocabularyName - The name of the vocabulary to use when processing a medical transcription job.

$sel:maxSpeakerLabels:MedicalTranscriptionSetting', medicalTranscriptionSetting_maxSpeakerLabels - The maximum number of speakers to identify in the input audio. If there are more speakers in the audio than this number, multiple speakers are identified as a single speaker. If you specify the MaxSpeakerLabels field, you must set the ShowSpeakerLabels field to true.

medicalTranscriptionSetting_showAlternatives :: Lens' MedicalTranscriptionSetting (Maybe Bool)

Determines whether alternative transcripts are generated along with the transcript that has the highest confidence. If you set ShowAlternatives field to true, you must also set the maximum number of alternatives to return in the MaxAlternatives field.

medicalTranscriptionSetting_channelIdentification :: Lens' MedicalTranscriptionSetting (Maybe Bool)

Instructs Amazon Transcribe Medical to process each audio channel separately and then merge the transcription output of each channel into a single transcription.

Amazon Transcribe Medical also produces a transcription of each item detected on an audio channel, including the start time and end time of the item and alternative transcriptions of item. The alternative transcriptions also come with confidence scores provided by Amazon Transcribe Medical.

You can't set both ShowSpeakerLabels and ChannelIdentification in the same request. If you set both, your request returns a BadRequestException

medicalTranscriptionSetting_maxAlternatives :: Lens' MedicalTranscriptionSetting (Maybe Natural)

The maximum number of alternatives that you tell the service to return. If you specify the MaxAlternatives field, you must set the ShowAlternatives field to true.

medicalTranscriptionSetting_showSpeakerLabels :: Lens' MedicalTranscriptionSetting (Maybe Bool)

Determines whether the transcription job uses speaker recognition to identify different speakers in the input audio. Speaker recognition labels individual speakers in the audio file. If you set the ShowSpeakerLabels field to true, you must also set the maximum number of speaker labels in the MaxSpeakerLabels field.

You can't set both ShowSpeakerLabels and ChannelIdentification in the same request. If you set both, your request returns a BadRequestException.

medicalTranscriptionSetting_vocabularyName :: Lens' MedicalTranscriptionSetting (Maybe Text)

The name of the vocabulary to use when processing a medical transcription job.

medicalTranscriptionSetting_maxSpeakerLabels :: Lens' MedicalTranscriptionSetting (Maybe Natural)

The maximum number of speakers to identify in the input audio. If there are more speakers in the audio than this number, multiple speakers are identified as a single speaker. If you specify the MaxSpeakerLabels field, you must set the ShowSpeakerLabels field to true.