Network.AWS.Transcribe.Types.Rule

Description

Documentation

data Rule

A condition in the call between the customer and the agent that you want to filter for.

See: newRule smart constructor.

Constructors

Rule'

Fields

sentimentFilter :: Maybe SentimentFilter

A condition that is applied to a particular customer sentiment.

interruptionFilter :: Maybe InterruptionFilter

A condition for a time period when either the customer or agent was interrupting the other person.

transcriptFilter :: Maybe TranscriptFilter

A condition that catches particular words or phrases based on a exact match. For example, if you set the phrase "I want to speak to the manager", only that exact phrase will be returned.

nonTalkTimeFilter :: Maybe NonTalkTimeFilter

A condition for a time period when neither the customer nor the agent was talking.

newRule :: Rule

Create a value of Rule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sentimentFilter:Rule', rule_sentimentFilter - A condition that is applied to a particular customer sentiment.

$sel:interruptionFilter:Rule', rule_interruptionFilter - A condition for a time period when either the customer or agent was interrupting the other person.

$sel:transcriptFilter:Rule', rule_transcriptFilter - A condition that catches particular words or phrases based on a exact match. For example, if you set the phrase "I want to speak to the manager", only that exact phrase will be returned.

$sel:nonTalkTimeFilter:Rule', rule_nonTalkTimeFilter - A condition for a time period when neither the customer nor the agent was talking.

rule_sentimentFilter :: Lens' Rule (Maybe SentimentFilter)

A condition that is applied to a particular customer sentiment.

rule_interruptionFilter :: Lens' Rule (Maybe InterruptionFilter)

A condition for a time period when either the customer or agent was interrupting the other person.

rule_transcriptFilter :: Lens' Rule (Maybe TranscriptFilter)

A condition that catches particular words or phrases based on a exact match. For example, if you set the phrase "I want to speak to the manager", only that exact phrase will be returned.

rule_nonTalkTimeFilter :: Lens' Rule (Maybe NonTalkTimeFilter)

A condition for a time period when neither the customer nor the agent was talking.