Network.AWS.Transcribe.Types.SentimentFilter

Description

Documentation

data SentimentFilter

An object that enables you to specify a particular customer or agent sentiment. If at least 50 percent of the conversation turns (the back-and-forth between two speakers) in a specified time period match the specified sentiment, Amazon Transcribe will consider the sentiment a match.

See: newSentimentFilter smart constructor.

Constructors

SentimentFilter'

Fields

relativeTimeRange :: Maybe RelativeTimeRange

The time range, set in percentages, that correspond to proportion of the call.

negate :: Maybe Bool

Set to TRUE to look for sentiments that weren't specified in the request.

participantRole :: Maybe ParticipantRole

A value that determines whether the sentiment belongs to the customer or the agent.

absoluteTimeRange :: Maybe AbsoluteTimeRange

The time range, measured in seconds, of the sentiment.

sentiments :: NonEmpty SentimentValue

An array that enables you to specify sentiments for the customer or agent. You can specify one or more values.

newSentimentFilter

Arguments

:: NonEmpty SentimentValue
$sel:sentiments:SentimentFilter'

-> SentimentFilter

Create a value of SentimentFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:relativeTimeRange:SentimentFilter', sentimentFilter_relativeTimeRange - The time range, set in percentages, that correspond to proportion of the call.

$sel:negate:SentimentFilter', sentimentFilter_negate - Set to TRUE to look for sentiments that weren't specified in the request.

$sel:participantRole:SentimentFilter', sentimentFilter_participantRole - A value that determines whether the sentiment belongs to the customer or the agent.

$sel:absoluteTimeRange:SentimentFilter', sentimentFilter_absoluteTimeRange - The time range, measured in seconds, of the sentiment.

$sel:sentiments:SentimentFilter', sentimentFilter_sentiments - An array that enables you to specify sentiments for the customer or agent. You can specify one or more values.

sentimentFilter_relativeTimeRange :: Lens' SentimentFilter (Maybe RelativeTimeRange)

The time range, set in percentages, that correspond to proportion of the call.

sentimentFilter_negate :: Lens' SentimentFilter (Maybe Bool)

Set to TRUE to look for sentiments that weren't specified in the request.

sentimentFilter_participantRole :: Lens' SentimentFilter (Maybe ParticipantRole)

A value that determines whether the sentiment belongs to the customer or the agent.

sentimentFilter_absoluteTimeRange :: Lens' SentimentFilter (Maybe AbsoluteTimeRange)

The time range, measured in seconds, of the sentiment.

sentimentFilter_sentiments :: Lens' SentimentFilter (NonEmpty SentimentValue)

An array that enables you to specify sentiments for the customer or agent. You can specify one or more values.