Network.AWS.Transcribe.Types.Transcript

Description

Documentation

data Transcript

Identifies the location of a transcription.

See: newTranscript smart constructor.

Constructors

Transcript'

Fields

transcriptFileUri :: Maybe Text

The S3 object location of the transcript.

Use this URI to access the transcript. If you specified an S3 bucket in the OutputBucketName field when you created the job, this is the URI of that bucket. If you chose to store the transcript in Amazon Transcribe, this is a shareable URL that provides secure access to that location.

redactedTranscriptFileUri :: Maybe Text

The S3 object location of the redacted transcript.

Use this URI to access the redacted transcript. If you specified an S3 bucket in the OutputBucketName field when you created the job, this is the URI of that bucket. If you chose to store the transcript in Amazon Transcribe, this is a shareable URL that provides secure access to that location.

newTranscript :: Transcript

Create a value of Transcript with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transcriptFileUri:Transcript', transcript_transcriptFileUri - The S3 object location of the transcript.

Use this URI to access the transcript. If you specified an S3 bucket in the OutputBucketName field when you created the job, this is the URI of that bucket. If you chose to store the transcript in Amazon Transcribe, this is a shareable URL that provides secure access to that location.

$sel:redactedTranscriptFileUri:Transcript', transcript_redactedTranscriptFileUri - The S3 object location of the redacted transcript.

Use this URI to access the redacted transcript. If you specified an S3 bucket in the OutputBucketName field when you created the job, this is the URI of that bucket. If you chose to store the transcript in Amazon Transcribe, this is a shareable URL that provides secure access to that location.

transcript_transcriptFileUri :: Lens' Transcript (Maybe Text)

The S3 object location of the transcript.

Use this URI to access the transcript. If you specified an S3 bucket in the OutputBucketName field when you created the job, this is the URI of that bucket. If you chose to store the transcript in Amazon Transcribe, this is a shareable URL that provides secure access to that location.

transcript_redactedTranscriptFileUri :: Lens' Transcript (Maybe Text)

The S3 object location of the redacted transcript.

Use this URI to access the redacted transcript. If you specified an S3 bucket in the OutputBucketName field when you created the job, this is the URI of that bucket. If you chose to store the transcript in Amazon Transcribe, this is a shareable URL that provides secure access to that location.