Network.AWS.Transcribe.Types.TranscriptFilter

Description

Documentation

data TranscriptFilter

Matches the output of the transcription to either the specific phrases that you specify, or the intent of the phrases that you specify.

See: newTranscriptFilter smart constructor.

Constructors

TranscriptFilter'

Fields

relativeTimeRange :: Maybe RelativeTimeRange

An object that allows percentages to specify the proportion of the call where you would like to apply a filter. For example, you can specify the first half of the call. You can also specify the period of time between halfway through to three-quarters of the way through the call. Because the length of conversation can vary between calls, you can apply relative time ranges across all calls.

negate :: Maybe Bool

If TRUE, the rule that you specify is applied to everything except for the phrases that you specify.

participantRole :: Maybe ParticipantRole

Determines whether the customer or the agent is speaking the phrases that you've specified.

absoluteTimeRange :: Maybe AbsoluteTimeRange

A time range, set in seconds, between two points in the call.

transcriptFilterType :: TranscriptFilterType

Matches the phrase to the transcription output in a word for word fashion. For example, if you specify the phrase "I want to speak to the manager." Amazon Transcribe attempts to match that specific phrase to the transcription.

targets :: NonEmpty Text

The phrases that you're specifying for the transcript filter to match.

newTranscriptFilter

Arguments

:: TranscriptFilterType
$sel:transcriptFilterType:TranscriptFilter'

-> NonEmpty Text
$sel:targets:TranscriptFilter'

-> TranscriptFilter

Create a value of TranscriptFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:relativeTimeRange:TranscriptFilter', transcriptFilter_relativeTimeRange - An object that allows percentages to specify the proportion of the call where you would like to apply a filter. For example, you can specify the first half of the call. You can also specify the period of time between halfway through to three-quarters of the way through the call. Because the length of conversation can vary between calls, you can apply relative time ranges across all calls.

$sel:negate:TranscriptFilter', transcriptFilter_negate - If TRUE, the rule that you specify is applied to everything except for the phrases that you specify.

$sel:participantRole:TranscriptFilter', transcriptFilter_participantRole - Determines whether the customer or the agent is speaking the phrases that you've specified.

$sel:absoluteTimeRange:TranscriptFilter', transcriptFilter_absoluteTimeRange - A time range, set in seconds, between two points in the call.

$sel:transcriptFilterType:TranscriptFilter', transcriptFilter_transcriptFilterType - Matches the phrase to the transcription output in a word for word fashion. For example, if you specify the phrase "I want to speak to the manager." Amazon Transcribe attempts to match that specific phrase to the transcription.

$sel:targets:TranscriptFilter', transcriptFilter_targets - The phrases that you're specifying for the transcript filter to match.

transcriptFilter_relativeTimeRange :: Lens' TranscriptFilter (Maybe RelativeTimeRange)

An object that allows percentages to specify the proportion of the call where you would like to apply a filter. For example, you can specify the first half of the call. You can also specify the period of time between halfway through to three-quarters of the way through the call. Because the length of conversation can vary between calls, you can apply relative time ranges across all calls.

transcriptFilter_negate :: Lens' TranscriptFilter (Maybe Bool)

If TRUE, the rule that you specify is applied to everything except for the phrases that you specify.

transcriptFilter_participantRole :: Lens' TranscriptFilter (Maybe ParticipantRole)

Determines whether the customer or the agent is speaking the phrases that you've specified.

transcriptFilter_absoluteTimeRange :: Lens' TranscriptFilter (Maybe AbsoluteTimeRange)

A time range, set in seconds, between two points in the call.

transcriptFilter_transcriptFilterType :: Lens' TranscriptFilter TranscriptFilterType

Matches the phrase to the transcription output in a word for word fashion. For example, if you specify the phrase "I want to speak to the manager." Amazon Transcribe attempts to match that specific phrase to the transcription.

transcriptFilter_targets :: Lens' TranscriptFilter (NonEmpty Text)

The phrases that you're specifying for the transcript filter to match.