Network.AWS.Transcribe.Types.TranscriptionJobSummary

Description

Documentation

data TranscriptionJobSummary

Provides a summary of information about a transcription job.

See: newTranscriptionJobSummary smart constructor.

Constructors

TranscriptionJobSummary'

Fields

languageCode :: Maybe LanguageCode

The language code for the input speech.

contentRedaction :: Maybe ContentRedaction

The content redaction settings of the transcription job.

creationTime :: Maybe POSIX

A timestamp that shows when the job was created.

completionTime :: Maybe POSIX

A timestamp that shows when the job was completed.

identifyLanguage :: Maybe Bool

Whether automatic language identification was enabled for a transcription job.

transcriptionJobName :: Maybe Text

The name of the transcription job.

transcriptionJobStatus :: Maybe TranscriptionJobStatus

The status of the transcription job. When the status is COMPLETED, use the GetTranscriptionJob operation to get the results of the transcription.

startTime :: Maybe POSIX

A timestamp that shows when the job started processing.

outputLocationType :: Maybe OutputLocationType

Indicates the location of the output of the transcription job.

If the value is CUSTOMER_BUCKET then the location is the S3 bucket specified in the outputBucketName field when the transcription job was started with the StartTranscriptionJob operation.

If the value is SERVICE_BUCKET then the output is stored by Amazon Transcribe and can be retrieved using the URI in the GetTranscriptionJob response's TranscriptFileUri field.

modelSettings :: Maybe ModelSettings

identifiedLanguageScore :: Maybe Double

A value between zero and one that Amazon Transcribe assigned to the language it identified in the source audio. A higher score indicates that Amazon Transcribe is more confident in the language it identified.

failureReason :: Maybe Text

If the TranscriptionJobStatus field is FAILED, a description of the error.

newTranscriptionJobSummary :: TranscriptionJobSummary

Create a value of TranscriptionJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:TranscriptionJobSummary', transcriptionJobSummary_languageCode - The language code for the input speech.

$sel:contentRedaction:TranscriptionJobSummary', transcriptionJobSummary_contentRedaction - The content redaction settings of the transcription job.

$sel:creationTime:TranscriptionJobSummary', transcriptionJobSummary_creationTime - A timestamp that shows when the job was created.

$sel:completionTime:TranscriptionJobSummary', transcriptionJobSummary_completionTime - A timestamp that shows when the job was completed.

$sel:identifyLanguage:TranscriptionJobSummary', transcriptionJobSummary_identifyLanguage - Whether automatic language identification was enabled for a transcription job.

$sel:transcriptionJobName:TranscriptionJobSummary', transcriptionJobSummary_transcriptionJobName - The name of the transcription job.

$sel:transcriptionJobStatus:TranscriptionJobSummary', transcriptionJobSummary_transcriptionJobStatus - The status of the transcription job. When the status is COMPLETED, use the GetTranscriptionJob operation to get the results of the transcription.

$sel:startTime:TranscriptionJobSummary', transcriptionJobSummary_startTime - A timestamp that shows when the job started processing.

$sel:outputLocationType:TranscriptionJobSummary', transcriptionJobSummary_outputLocationType - Indicates the location of the output of the transcription job.

If the value is CUSTOMER_BUCKET then the location is the S3 bucket specified in the outputBucketName field when the transcription job was started with the StartTranscriptionJob operation.

If the value is SERVICE_BUCKET then the output is stored by Amazon Transcribe and can be retrieved using the URI in the GetTranscriptionJob response's TranscriptFileUri field.

$sel:modelSettings:TranscriptionJobSummary', transcriptionJobSummary_modelSettings - Undocumented member.

$sel:identifiedLanguageScore:TranscriptionJobSummary', transcriptionJobSummary_identifiedLanguageScore - A value between zero and one that Amazon Transcribe assigned to the language it identified in the source audio. A higher score indicates that Amazon Transcribe is more confident in the language it identified.

$sel:failureReason:TranscriptionJobSummary', transcriptionJobSummary_failureReason - If the TranscriptionJobStatus field is FAILED, a description of the error.

transcriptionJobSummary_languageCode :: Lens' TranscriptionJobSummary (Maybe LanguageCode)

The language code for the input speech.

transcriptionJobSummary_contentRedaction :: Lens' TranscriptionJobSummary (Maybe ContentRedaction)

The content redaction settings of the transcription job.

transcriptionJobSummary_creationTime :: Lens' TranscriptionJobSummary (Maybe UTCTime)

A timestamp that shows when the job was created.

transcriptionJobSummary_completionTime :: Lens' TranscriptionJobSummary (Maybe UTCTime)

A timestamp that shows when the job was completed.

transcriptionJobSummary_identifyLanguage :: Lens' TranscriptionJobSummary (Maybe Bool)

Whether automatic language identification was enabled for a transcription job.

transcriptionJobSummary_transcriptionJobName :: Lens' TranscriptionJobSummary (Maybe Text)

The name of the transcription job.

transcriptionJobSummary_transcriptionJobStatus :: Lens' TranscriptionJobSummary (Maybe TranscriptionJobStatus)

The status of the transcription job. When the status is COMPLETED, use the GetTranscriptionJob operation to get the results of the transcription.

transcriptionJobSummary_startTime :: Lens' TranscriptionJobSummary (Maybe UTCTime)

A timestamp that shows when the job started processing.

transcriptionJobSummary_outputLocationType :: Lens' TranscriptionJobSummary (Maybe OutputLocationType)

Indicates the location of the output of the transcription job.

If the value is CUSTOMER_BUCKET then the location is the S3 bucket specified in the outputBucketName field when the transcription job was started with the StartTranscriptionJob operation.

If the value is SERVICE_BUCKET then the output is stored by Amazon Transcribe and can be retrieved using the URI in the GetTranscriptionJob response's TranscriptFileUri field.

transcriptionJobSummary_modelSettings :: Lens' TranscriptionJobSummary (Maybe ModelSettings)

Undocumented member.

transcriptionJobSummary_identifiedLanguageScore :: Lens' TranscriptionJobSummary (Maybe Double)

A value between zero and one that Amazon Transcribe assigned to the language it identified in the source audio. A higher score indicates that Amazon Transcribe is more confident in the language it identified.

transcriptionJobSummary_failureReason :: Lens' TranscriptionJobSummary (Maybe Text)

If the TranscriptionJobStatus field is FAILED, a description of the error.