Network.AWS.Transcribe.Types.VocabularyInfo

Description

Documentation

data VocabularyInfo

Provides information about a custom vocabulary.

See: newVocabularyInfo smart constructor.

Constructors

VocabularyInfo'

Fields

languageCode :: Maybe LanguageCode

The language code of the vocabulary entries.

lastModifiedTime :: Maybe POSIX

The date and time that the vocabulary was last modified.

vocabularyName :: Maybe Text

The name of the vocabulary.

vocabularyState :: Maybe VocabularyState

The processing state of the vocabulary. If the state is READY you can use the vocabulary in a StartTranscriptionJob request.

newVocabularyInfo :: VocabularyInfo

Create a value of VocabularyInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:VocabularyInfo', vocabularyInfo_languageCode - The language code of the vocabulary entries.

$sel:lastModifiedTime:VocabularyInfo', vocabularyInfo_lastModifiedTime - The date and time that the vocabulary was last modified.

$sel:vocabularyName:VocabularyInfo', vocabularyInfo_vocabularyName - The name of the vocabulary.

$sel:vocabularyState:VocabularyInfo', vocabularyInfo_vocabularyState - The processing state of the vocabulary. If the state is READY you can use the vocabulary in a StartTranscriptionJob request.

vocabularyInfo_languageCode :: Lens' VocabularyInfo (Maybe LanguageCode)

The language code of the vocabulary entries.

vocabularyInfo_lastModifiedTime :: Lens' VocabularyInfo (Maybe UTCTime)

The date and time that the vocabulary was last modified.

vocabularyInfo_vocabularyName :: Lens' VocabularyInfo (Maybe Text)

The name of the vocabulary.

vocabularyInfo_vocabularyState :: Lens' VocabularyInfo (Maybe VocabularyState)

The processing state of the vocabulary. If the state is READY you can use the vocabulary in a StartTranscriptionJob request.