Network.AWS.Transcribe.UpdateMedicalVocabulary

Description

Updates a vocabulary with new values that you provide in a different text file from the one you used to create the vocabulary. The UpdateMedicalVocabulary operation overwrites all of the existing information with the values that you provide in the request.

Creating a Request

data UpdateMedicalVocabulary

See: newUpdateMedicalVocabulary smart constructor.

Constructors

UpdateMedicalVocabulary'

Fields

vocabularyFileUri :: Maybe Text

The location in Amazon S3 of the text file that contains your custom vocabulary. The URI must be in the same Amazon Web Services Region as the resource that you are calling. The following is the format for a URI:

 https://s3.<aws-region>.amazonaws.com/<bucket-name>/<keyprefix>/<objectkey>

For example:

https://s3.us-east-1.amazonaws.com/AWSDOC-EXAMPLE-BUCKET/vocab.txt

For more information about Amazon S3 object names, see Object Keys in the Amazon S3 Developer Guide.

For more information about custom vocabularies in Amazon Transcribe Medical, see Medical Custom Vocabularies.

vocabularyName :: Text

The name of the vocabulary to update. The name is case sensitive. If you try to update a vocabulary with the same name as a vocabulary you've already made, you get a ConflictException error.

languageCode :: LanguageCode

The language code of the language used for the entries in the updated vocabulary. US English (en-US) is the only valid language code in Amazon Transcribe Medical.

newUpdateMedicalVocabulary

Arguments

:: Text
$sel:vocabularyName:UpdateMedicalVocabulary'

-> LanguageCode
$sel:languageCode:UpdateMedicalVocabulary'

-> UpdateMedicalVocabulary

Create a value of UpdateMedicalVocabulary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vocabularyFileUri:UpdateMedicalVocabulary', updateMedicalVocabulary_vocabularyFileUri - The location in Amazon S3 of the text file that contains your custom vocabulary. The URI must be in the same Amazon Web Services Region as the resource that you are calling. The following is the format for a URI:

 https://s3.<aws-region>.amazonaws.com/<bucket-name>/<keyprefix>/<objectkey>

For example:

https://s3.us-east-1.amazonaws.com/AWSDOC-EXAMPLE-BUCKET/vocab.txt

For more information about Amazon S3 object names, see Object Keys in the Amazon S3 Developer Guide.

For more information about custom vocabularies in Amazon Transcribe Medical, see Medical Custom Vocabularies.

$sel:vocabularyName:UpdateMedicalVocabulary', updateMedicalVocabulary_vocabularyName - The name of the vocabulary to update. The name is case sensitive. If you try to update a vocabulary with the same name as a vocabulary you've already made, you get a ConflictException error.

$sel:languageCode:UpdateMedicalVocabulary', updateMedicalVocabulary_languageCode - The language code of the language used for the entries in the updated vocabulary. US English (en-US) is the only valid language code in Amazon Transcribe Medical.

Request Lenses

updateMedicalVocabulary_vocabularyFileUri :: Lens' UpdateMedicalVocabulary (Maybe Text)

The location in Amazon S3 of the text file that contains your custom vocabulary. The URI must be in the same Amazon Web Services Region as the resource that you are calling. The following is the format for a URI:

 https://s3.<aws-region>.amazonaws.com/<bucket-name>/<keyprefix>/<objectkey>

For example:

https://s3.us-east-1.amazonaws.com/AWSDOC-EXAMPLE-BUCKET/vocab.txt

For more information about Amazon S3 object names, see Object Keys in the Amazon S3 Developer Guide.

For more information about custom vocabularies in Amazon Transcribe Medical, see Medical Custom Vocabularies.

updateMedicalVocabulary_vocabularyName :: Lens' UpdateMedicalVocabulary Text

The name of the vocabulary to update. The name is case sensitive. If you try to update a vocabulary with the same name as a vocabulary you've already made, you get a ConflictException error.

updateMedicalVocabulary_languageCode :: Lens' UpdateMedicalVocabulary LanguageCode

The language code of the language used for the entries in the updated vocabulary. US English (en-US) is the only valid language code in Amazon Transcribe Medical.

Destructuring the Response

data UpdateMedicalVocabularyResponse

See: newUpdateMedicalVocabularyResponse smart constructor.

Constructors

UpdateMedicalVocabularyResponse'

Fields

languageCode :: Maybe LanguageCode

The language code for the language of the text file used to update the custom vocabulary. US English (en-US) is the only language supported in Amazon Transcribe Medical.

lastModifiedTime :: Maybe POSIX

The date and time that the vocabulary was updated.

vocabularyName :: Maybe Text

The name of the updated vocabulary.

vocabularyState :: Maybe VocabularyState

The processing state of the update to the vocabulary. When the VocabularyState field is READY, the vocabulary is ready to be used in a StartMedicalTranscriptionJob request.

httpStatus :: Int

The response's http status code.

newUpdateMedicalVocabularyResponse

Arguments

:: Int
$sel:httpStatus:UpdateMedicalVocabularyResponse'

-> UpdateMedicalVocabularyResponse

Create a value of UpdateMedicalVocabularyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:UpdateMedicalVocabulary', updateMedicalVocabularyResponse_languageCode - The language code for the language of the text file used to update the custom vocabulary. US English (en-US) is the only language supported in Amazon Transcribe Medical.

$sel:lastModifiedTime:UpdateMedicalVocabularyResponse', updateMedicalVocabularyResponse_lastModifiedTime - The date and time that the vocabulary was updated.

$sel:vocabularyName:UpdateMedicalVocabulary', updateMedicalVocabularyResponse_vocabularyName - The name of the updated vocabulary.

$sel:vocabularyState:UpdateMedicalVocabularyResponse', updateMedicalVocabularyResponse_vocabularyState - The processing state of the update to the vocabulary. When the VocabularyState field is READY, the vocabulary is ready to be used in a StartMedicalTranscriptionJob request.

$sel:httpStatus:UpdateMedicalVocabularyResponse', updateMedicalVocabularyResponse_httpStatus - The response's http status code.

Response Lenses

updateMedicalVocabularyResponse_languageCode :: Lens' UpdateMedicalVocabularyResponse (Maybe LanguageCode)

The language code for the language of the text file used to update the custom vocabulary. US English (en-US) is the only language supported in Amazon Transcribe Medical.

updateMedicalVocabularyResponse_lastModifiedTime :: Lens' UpdateMedicalVocabularyResponse (Maybe UTCTime)

The date and time that the vocabulary was updated.

updateMedicalVocabularyResponse_vocabularyName :: Lens' UpdateMedicalVocabularyResponse (Maybe Text)

The name of the updated vocabulary.

updateMedicalVocabularyResponse_vocabularyState :: Lens' UpdateMedicalVocabularyResponse (Maybe VocabularyState)

The processing state of the update to the vocabulary. When the VocabularyState field is READY, the vocabulary is ready to be used in a StartMedicalTranscriptionJob request.

updateMedicalVocabularyResponse_httpStatus :: Lens' UpdateMedicalVocabularyResponse Int

The response's http status code.