Network.AWS.Transfer.DescribeUser

Description

Describes the user assigned to the specific file transfer protocol-enabled server, as identified by its ServerId property.

The response from this call returns the properties of the user associated with the ServerId value that was specified.

Creating a Request

data DescribeUser

See: newDescribeUser smart constructor.

Constructors

DescribeUser'

Fields

serverId :: Text

A system-assigned unique identifier for a server that has this user assigned.

userName :: Text

The name of the user assigned to one or more servers. User names are part of the sign-in credentials to use the Amazon Web Services Transfer Family service and perform file transfer tasks.

newDescribeUser

Arguments

:: Text
$sel:serverId:DescribeUser'

-> Text
$sel:userName:DescribeUser'

-> DescribeUser

Create a value of DescribeUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serverId:DescribeUser', describeUser_serverId - A system-assigned unique identifier for a server that has this user assigned.

$sel:userName:DescribeUser', describeUser_userName - The name of the user assigned to one or more servers. User names are part of the sign-in credentials to use the Amazon Web Services Transfer Family service and perform file transfer tasks.

Request Lenses

describeUser_serverId :: Lens' DescribeUser Text

A system-assigned unique identifier for a server that has this user assigned.

describeUser_userName :: Lens' DescribeUser Text

The name of the user assigned to one or more servers. User names are part of the sign-in credentials to use the Amazon Web Services Transfer Family service and perform file transfer tasks.

Destructuring the Response

data DescribeUserResponse

See: newDescribeUserResponse smart constructor.

Constructors

DescribeUserResponse'

Fields

httpStatus :: Int

The response's http status code.

serverId :: Text

A system-assigned unique identifier for a server that has this user assigned.

user :: DescribedUser

An array containing the properties of the user account for the ServerID value that you specified.

newDescribeUserResponse

Arguments

:: Int
$sel:httpStatus:DescribeUserResponse'

-> Text
$sel:serverId:DescribeUser'

-> DescribedUser
$sel:user:DescribeUserResponse'

-> DescribeUserResponse

Create a value of DescribeUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeUserResponse', describeUserResponse_httpStatus - The response's http status code.

$sel:serverId:DescribeUser', describeUserResponse_serverId - A system-assigned unique identifier for a server that has this user assigned.

$sel:user:DescribeUserResponse', describeUserResponse_user - An array containing the properties of the user account for the ServerID value that you specified.

Response Lenses

describeUserResponse_httpStatus :: Lens' DescribeUserResponse Int

The response's http status code.

describeUserResponse_serverId :: Lens' DescribeUserResponse Text

A system-assigned unique identifier for a server that has this user assigned.

describeUserResponse_user :: Lens' DescribeUserResponse DescribedUser

An array containing the properties of the user account for the ServerID value that you specified.