Network.AWS.Transfer.TestIdentityProvider

Description

If the IdentityProviderType of a file transfer protocol-enabled server is AWS_DIRECTORY_SERVICE or API_Gateway, tests whether your identity provider is set up successfully. We highly recommend that you call this operation to test your authentication method as soon as you create your server. By doing so, you can troubleshoot issues with the identity provider integration to ensure that your users can successfully use the service.

The ServerId and UserName parameters are required. The ServerProtocol, SourceIp, and UserPassword are all optional.

You cannot use TestIdentityProvider if the IdentityProviderType of your server is SERVICE_MANAGED.

If you provide any incorrect values for any parameters, the Response field is empty.

If you provide a server ID for a server that uses service-managed users, you get an error:

 An error occurred (InvalidRequestException) when calling the TestIdentityProvider operation: s-server-ID not configured for external auth

If you enter a Server ID for the --server-id parameter that does not identify an actual Transfer server, you receive the following error:

An error occurred (ResourceNotFoundException) when calling the TestIdentityProvider operation: Unknown server

Creating a Request

data TestIdentityProvider

See: newTestIdentityProvider smart constructor.

Constructors

TestIdentityProvider'

Fields

serverProtocol :: Maybe Protocol

The type of file transfer protocol to be tested.

The available protocols are:

Secure Shell (SSH) File Transfer Protocol (SFTP)

File Transfer Protocol Secure (FTPS)

File Transfer Protocol (FTP)

userPassword :: Maybe (Sensitive Text)

The password of the user account to be tested.

sourceIp :: Maybe Text

The source IP address of the user account to be tested.

serverId :: Text

A system-assigned identifier for a specific server. That server's user authentication method is tested with a user name and password.

userName :: Text

The name of the user account to be tested.

newTestIdentityProvider

Arguments

:: Text
$sel:serverId:TestIdentityProvider'

-> Text
$sel:userName:TestIdentityProvider'

-> TestIdentityProvider

Create a value of TestIdentityProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serverProtocol:TestIdentityProvider', testIdentityProvider_serverProtocol - The type of file transfer protocol to be tested.

The available protocols are:

Secure Shell (SSH) File Transfer Protocol (SFTP)

File Transfer Protocol Secure (FTPS)

File Transfer Protocol (FTP)

$sel:userPassword:TestIdentityProvider', testIdentityProvider_userPassword - The password of the user account to be tested.

$sel:sourceIp:TestIdentityProvider', testIdentityProvider_sourceIp - The source IP address of the user account to be tested.

$sel:serverId:TestIdentityProvider', testIdentityProvider_serverId - A system-assigned identifier for a specific server. That server's user authentication method is tested with a user name and password.

$sel:userName:TestIdentityProvider', testIdentityProvider_userName - The name of the user account to be tested.

Request Lenses

testIdentityProvider_serverProtocol :: Lens' TestIdentityProvider (Maybe Protocol)

The type of file transfer protocol to be tested.

The available protocols are:

Secure Shell (SSH) File Transfer Protocol (SFTP)

File Transfer Protocol Secure (FTPS)

File Transfer Protocol (FTP)

testIdentityProvider_userPassword :: Lens' TestIdentityProvider (Maybe Text)

The password of the user account to be tested.

testIdentityProvider_sourceIp :: Lens' TestIdentityProvider (Maybe Text)

The source IP address of the user account to be tested.

testIdentityProvider_serverId :: Lens' TestIdentityProvider Text

A system-assigned identifier for a specific server. That server's user authentication method is tested with a user name and password.

testIdentityProvider_userName :: Lens' TestIdentityProvider Text

The name of the user account to be tested.

Destructuring the Response

data TestIdentityProviderResponse

See: newTestIdentityProviderResponse smart constructor.

Constructors

TestIdentityProviderResponse'

Fields

response :: Maybe Text

The response that is returned from your API Gateway.

message :: Maybe Text

A message that indicates whether the test was successful or not.

httpStatus :: Int

The response's http status code.

statusCode :: Int

The HTTP status code that is the response from your API Gateway.

url :: Text

The endpoint of the service used to authenticate a user.

newTestIdentityProviderResponse

Arguments

:: Int
$sel:httpStatus:TestIdentityProviderResponse'

-> Int
$sel:statusCode:TestIdentityProviderResponse'

-> Text
$sel:url:TestIdentityProviderResponse'

-> TestIdentityProviderResponse

Create a value of TestIdentityProviderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:response:TestIdentityProviderResponse', testIdentityProviderResponse_response - The response that is returned from your API Gateway.

$sel:message:TestIdentityProviderResponse', testIdentityProviderResponse_message - A message that indicates whether the test was successful or not.

$sel:httpStatus:TestIdentityProviderResponse', testIdentityProviderResponse_httpStatus - The response's http status code.

$sel:statusCode:TestIdentityProviderResponse', testIdentityProviderResponse_statusCode - The HTTP status code that is the response from your API Gateway.

$sel:url:TestIdentityProviderResponse', testIdentityProviderResponse_url - The endpoint of the service used to authenticate a user.

Response Lenses

testIdentityProviderResponse_response :: Lens' TestIdentityProviderResponse (Maybe Text)

The response that is returned from your API Gateway.

testIdentityProviderResponse_message :: Lens' TestIdentityProviderResponse (Maybe Text)

A message that indicates whether the test was successful or not.

testIdentityProviderResponse_httpStatus :: Lens' TestIdentityProviderResponse Int

The response's http status code.

testIdentityProviderResponse_statusCode :: Lens' TestIdentityProviderResponse Int

The HTTP status code that is the response from your API Gateway.

testIdentityProviderResponse_url :: Lens' TestIdentityProviderResponse Text

The endpoint of the service used to authenticate a user.