Network.AWS.Transfer.Types.DescribedExecution

Description

Documentation

data DescribedExecution

The details for an execution object.

See: newDescribedExecution smart constructor.

Constructors

DescribedExecution'

Fields

status :: Maybe ExecutionStatus

The status is one of the execution. Can be in progress, completed, exception encountered, or handling the exception.

executionId :: Maybe Text

A unique identifier for the execution of a workflow.

results :: Maybe ExecutionResults

A structure that describes the execution results. This includes a list of the steps along with the details of each step, error type and message (if any), and the OnExceptionSteps structure.

initialFileLocation :: Maybe FileLocation

A structure that describes the Amazon S3 or EFS file location. This is the file location when the execution begins: if the file is being copied, this is the initial (as opposed to destination) file location.

posixProfile :: Maybe PosixProfile

serviceMetadata :: Maybe ServiceMetadata

A container object for the session details associated with a workflow.

loggingConfiguration :: Maybe LoggingConfiguration

The IAM logging role associated with the execution.

executionRole :: Maybe Text

The IAM role associated with the execution.

newDescribedExecution :: DescribedExecution

Create a value of DescribedExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribedExecution', describedExecution_status - The status is one of the execution. Can be in progress, completed, exception encountered, or handling the exception.

$sel:executionId:DescribedExecution', describedExecution_executionId - A unique identifier for the execution of a workflow.

$sel:results:DescribedExecution', describedExecution_results - A structure that describes the execution results. This includes a list of the steps along with the details of each step, error type and message (if any), and the OnExceptionSteps structure.

$sel:initialFileLocation:DescribedExecution', describedExecution_initialFileLocation - A structure that describes the Amazon S3 or EFS file location. This is the file location when the execution begins: if the file is being copied, this is the initial (as opposed to destination) file location.

$sel:posixProfile:DescribedExecution', describedExecution_posixProfile - Undocumented member.

$sel:serviceMetadata:DescribedExecution', describedExecution_serviceMetadata - A container object for the session details associated with a workflow.

$sel:loggingConfiguration:DescribedExecution', describedExecution_loggingConfiguration - The IAM logging role associated with the execution.

$sel:executionRole:DescribedExecution', describedExecution_executionRole - The IAM role associated with the execution.

describedExecution_status :: Lens' DescribedExecution (Maybe ExecutionStatus)

The status is one of the execution. Can be in progress, completed, exception encountered, or handling the exception.

describedExecution_executionId :: Lens' DescribedExecution (Maybe Text)

A unique identifier for the execution of a workflow.

describedExecution_results :: Lens' DescribedExecution (Maybe ExecutionResults)

A structure that describes the execution results. This includes a list of the steps along with the details of each step, error type and message (if any), and the OnExceptionSteps structure.

describedExecution_initialFileLocation :: Lens' DescribedExecution (Maybe FileLocation)

A structure that describes the Amazon S3 or EFS file location. This is the file location when the execution begins: if the file is being copied, this is the initial (as opposed to destination) file location.

describedExecution_posixProfile :: Lens' DescribedExecution (Maybe PosixProfile)

Undocumented member.

describedExecution_serviceMetadata :: Lens' DescribedExecution (Maybe ServiceMetadata)

A container object for the session details associated with a workflow.

describedExecution_loggingConfiguration :: Lens' DescribedExecution (Maybe LoggingConfiguration)

The IAM logging role associated with the execution.

describedExecution_executionRole :: Lens' DescribedExecution (Maybe Text)

The IAM role associated with the execution.