Network.AWS.Transfer.Types.DescribedServer

Description

Documentation

data DescribedServer

Describes the properties of a file transfer protocol-enabled server that was specified.

See: newDescribedServer smart constructor.

Constructors

DescribedServer'

Fields

protocolDetails :: Maybe ProtocolDetails

The protocol settings that are configured for your server.

Use the PassiveIp parameter to indicate passive mode. Enter a single dotted-quad IPv4 address, such as the external IP address of a firewall, router, or load balancer.

loggingRole :: Maybe Text

Specifies the Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that allows a server to turn on Amazon CloudWatch logging for Amazon S3 or Amazon EFS events. When set, user activity can be viewed in your CloudWatch logs.

state :: Maybe State

Specifies the condition of a server for the server that was described. A value of ONLINE indicates that the server can accept jobs and transfer files. A State value of OFFLINE means that the server cannot perform file transfer operations.

The states of STARTING and STOPPING indicate that the server is in an intermediate state, either not fully able to respond, or not fully offline. The values of START_FAILED or STOP_FAILED can indicate an error condition.

identityProviderType :: Maybe IdentityProviderType

Specifies the mode of authentication for a server. The default value is SERVICE_MANAGED, which allows you to store and access user credentials within the Amazon Web Services Transfer Family service.

Use AWS_DIRECTORY_SERVICE to provide access to Active Directory groups in Amazon Web Services Managed Active Directory or Microsoft Active Directory in your on-premises environment or in Amazon Web Services using AD Connectors. This option also requires you to provide a Directory ID using the IdentityProviderDetails parameter.

Use the API_GATEWAY value to integrate with an identity provider of your choosing. The API_GATEWAY setting requires you to provide an API Gateway endpoint URL to call for authentication using the IdentityProviderDetails parameter.

protocols :: Maybe (NonEmpty Protocol)

Specifies the file transfer protocol or protocols over which your file transfer protocol client can connect to your server's endpoint. The available protocols are:

SFTP (Secure Shell (SSH) File Transfer Protocol): File transfer over SSH

FTPS (File Transfer Protocol Secure): File transfer with TLS encryption

FTP (File Transfer Protocol): Unencrypted file transfer

serverId :: Maybe Text

Specifies the unique system-assigned identifier for a server that you instantiate.

domain :: Maybe Domain

Specifies the domain of the storage system that is used for file transfers.

endpointType :: Maybe EndpointType

Defines the type of endpoint that your server is connected to. If your server is connected to a VPC endpoint, your server isn't accessible over the public internet.

securityPolicyName :: Maybe Text

Specifies the name of the security policy that is attached to the server.

hostKeyFingerprint :: Maybe Text

Specifies the Base64-encoded SHA256 fingerprint of the server's host key. This value is equivalent to the output of the ssh-keygen -l -f my-new-server-key command.

userCount :: Maybe Int

Specifies the number of users that are assigned to a server you specified with the ServerId.

certificate :: Maybe Text

Specifies the ARN of the Amazon Web ServicesCertificate Manager (ACM) certificate. Required when Protocols is set to FTPS.

identityProviderDetails :: Maybe IdentityProviderDetails

Specifies information to call a customer-supplied authentication API. This field is not populated when the IdentityProviderType of a server is AWS_DIRECTORY_SERVICE or SERVICE_MANAGED.

workflowDetails :: Maybe WorkflowDetails

Specifies the workflow ID for the workflow to assign and the execution role used for executing the workflow.

tags :: Maybe (NonEmpty Tag)

Specifies the key-value pairs that you can use to search for and group servers that were assigned to the server that was described.

endpointDetails :: Maybe EndpointDetails

The virtual private cloud (VPC) endpoint settings that are configured for your server. When you host your endpoint within your VPC, you can make it accessible only to resources within your VPC, or you can attach Elastic IP addresses and make it accessible to clients over the internet. Your VPC's default security groups are automatically assigned to your endpoint.

arn :: Text

Specifies the unique Amazon Resource Name (ARN) of the server.

newDescribedServer

Arguments

:: Text
$sel:arn:DescribedServer'

-> DescribedServer

Create a value of DescribedServer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocolDetails:DescribedServer', describedServer_protocolDetails - The protocol settings that are configured for your server.

Use the PassiveIp parameter to indicate passive mode. Enter a single dotted-quad IPv4 address, such as the external IP address of a firewall, router, or load balancer.

$sel:loggingRole:DescribedServer', describedServer_loggingRole - Specifies the Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that allows a server to turn on Amazon CloudWatch logging for Amazon S3 or Amazon EFS events. When set, user activity can be viewed in your CloudWatch logs.

$sel:state:DescribedServer', describedServer_state - Specifies the condition of a server for the server that was described. A value of ONLINE indicates that the server can accept jobs and transfer files. A State value of OFFLINE means that the server cannot perform file transfer operations.

The states of STARTING and STOPPING indicate that the server is in an intermediate state, either not fully able to respond, or not fully offline. The values of START_FAILED or STOP_FAILED can indicate an error condition.

$sel:identityProviderType:DescribedServer', describedServer_identityProviderType - Specifies the mode of authentication for a server. The default value is SERVICE_MANAGED, which allows you to store and access user credentials within the Amazon Web Services Transfer Family service.

Use AWS_DIRECTORY_SERVICE to provide access to Active Directory groups in Amazon Web Services Managed Active Directory or Microsoft Active Directory in your on-premises environment or in Amazon Web Services using AD Connectors. This option also requires you to provide a Directory ID using the IdentityProviderDetails parameter.

Use the API_GATEWAY value to integrate with an identity provider of your choosing. The API_GATEWAY setting requires you to provide an API Gateway endpoint URL to call for authentication using the IdentityProviderDetails parameter.

$sel:protocols:DescribedServer', describedServer_protocols - Specifies the file transfer protocol or protocols over which your file transfer protocol client can connect to your server's endpoint. The available protocols are:

SFTP (Secure Shell (SSH) File Transfer Protocol): File transfer over SSH

FTPS (File Transfer Protocol Secure): File transfer with TLS encryption

FTP (File Transfer Protocol): Unencrypted file transfer

$sel:serverId:DescribedServer', describedServer_serverId - Specifies the unique system-assigned identifier for a server that you instantiate.

$sel:domain:DescribedServer', describedServer_domain - Specifies the domain of the storage system that is used for file transfers.

$sel:endpointType:DescribedServer', describedServer_endpointType - Defines the type of endpoint that your server is connected to. If your server is connected to a VPC endpoint, your server isn't accessible over the public internet.

$sel:securityPolicyName:DescribedServer', describedServer_securityPolicyName - Specifies the name of the security policy that is attached to the server.

$sel:hostKeyFingerprint:DescribedServer', describedServer_hostKeyFingerprint - Specifies the Base64-encoded SHA256 fingerprint of the server's host key. This value is equivalent to the output of the ssh-keygen -l -f my-new-server-key command.

$sel:userCount:DescribedServer', describedServer_userCount - Specifies the number of users that are assigned to a server you specified with the ServerId.

$sel:certificate:DescribedServer', describedServer_certificate - Specifies the ARN of the Amazon Web ServicesCertificate Manager (ACM) certificate. Required when Protocols is set to FTPS.

$sel:identityProviderDetails:DescribedServer', describedServer_identityProviderDetails - Specifies information to call a customer-supplied authentication API. This field is not populated when the IdentityProviderType of a server is AWS_DIRECTORY_SERVICE or SERVICE_MANAGED.

$sel:workflowDetails:DescribedServer', describedServer_workflowDetails - Specifies the workflow ID for the workflow to assign and the execution role used for executing the workflow.

$sel:tags:DescribedServer', describedServer_tags - Specifies the key-value pairs that you can use to search for and group servers that were assigned to the server that was described.

$sel:endpointDetails:DescribedServer', describedServer_endpointDetails - The virtual private cloud (VPC) endpoint settings that are configured for your server. When you host your endpoint within your VPC, you can make it accessible only to resources within your VPC, or you can attach Elastic IP addresses and make it accessible to clients over the internet. Your VPC's default security groups are automatically assigned to your endpoint.

$sel:arn:DescribedServer', describedServer_arn - Specifies the unique Amazon Resource Name (ARN) of the server.

describedServer_protocolDetails :: Lens' DescribedServer (Maybe ProtocolDetails)

The protocol settings that are configured for your server.

Use the PassiveIp parameter to indicate passive mode. Enter a single dotted-quad IPv4 address, such as the external IP address of a firewall, router, or load balancer.

describedServer_loggingRole :: Lens' DescribedServer (Maybe Text)

Specifies the Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that allows a server to turn on Amazon CloudWatch logging for Amazon S3 or Amazon EFS events. When set, user activity can be viewed in your CloudWatch logs.

describedServer_state :: Lens' DescribedServer (Maybe State)

Specifies the condition of a server for the server that was described. A value of ONLINE indicates that the server can accept jobs and transfer files. A State value of OFFLINE means that the server cannot perform file transfer operations.

The states of STARTING and STOPPING indicate that the server is in an intermediate state, either not fully able to respond, or not fully offline. The values of START_FAILED or STOP_FAILED can indicate an error condition.

describedServer_identityProviderType :: Lens' DescribedServer (Maybe IdentityProviderType)

Specifies the mode of authentication for a server. The default value is SERVICE_MANAGED, which allows you to store and access user credentials within the Amazon Web Services Transfer Family service.

Use AWS_DIRECTORY_SERVICE to provide access to Active Directory groups in Amazon Web Services Managed Active Directory or Microsoft Active Directory in your on-premises environment or in Amazon Web Services using AD Connectors. This option also requires you to provide a Directory ID using the IdentityProviderDetails parameter.

Use the API_GATEWAY value to integrate with an identity provider of your choosing. The API_GATEWAY setting requires you to provide an API Gateway endpoint URL to call for authentication using the IdentityProviderDetails parameter.

describedServer_protocols :: Lens' DescribedServer (Maybe (NonEmpty Protocol))

Specifies the file transfer protocol or protocols over which your file transfer protocol client can connect to your server's endpoint. The available protocols are:

SFTP (Secure Shell (SSH) File Transfer Protocol): File transfer over SSH

FTPS (File Transfer Protocol Secure): File transfer with TLS encryption

FTP (File Transfer Protocol): Unencrypted file transfer

describedServer_serverId :: Lens' DescribedServer (Maybe Text)

Specifies the unique system-assigned identifier for a server that you instantiate.

describedServer_domain :: Lens' DescribedServer (Maybe Domain)

Specifies the domain of the storage system that is used for file transfers.

describedServer_endpointType :: Lens' DescribedServer (Maybe EndpointType)

Defines the type of endpoint that your server is connected to. If your server is connected to a VPC endpoint, your server isn't accessible over the public internet.

describedServer_securityPolicyName :: Lens' DescribedServer (Maybe Text)

Specifies the name of the security policy that is attached to the server.

describedServer_hostKeyFingerprint :: Lens' DescribedServer (Maybe Text)

Specifies the Base64-encoded SHA256 fingerprint of the server's host key. This value is equivalent to the output of the ssh-keygen -l -f my-new-server-key command.

describedServer_userCount :: Lens' DescribedServer (Maybe Int)

Specifies the number of users that are assigned to a server you specified with the ServerId.

describedServer_certificate :: Lens' DescribedServer (Maybe Text)

Specifies the ARN of the Amazon Web ServicesCertificate Manager (ACM) certificate. Required when Protocols is set to FTPS.

describedServer_identityProviderDetails :: Lens' DescribedServer (Maybe IdentityProviderDetails)

Specifies information to call a customer-supplied authentication API. This field is not populated when the IdentityProviderType of a server is AWS_DIRECTORY_SERVICE or SERVICE_MANAGED.

describedServer_workflowDetails :: Lens' DescribedServer (Maybe WorkflowDetails)

Specifies the workflow ID for the workflow to assign and the execution role used for executing the workflow.

describedServer_tags :: Lens' DescribedServer (Maybe (NonEmpty Tag))

Specifies the key-value pairs that you can use to search for and group servers that were assigned to the server that was described.

describedServer_endpointDetails :: Lens' DescribedServer (Maybe EndpointDetails)

The virtual private cloud (VPC) endpoint settings that are configured for your server. When you host your endpoint within your VPC, you can make it accessible only to resources within your VPC, or you can attach Elastic IP addresses and make it accessible to clients over the internet. Your VPC's default security groups are automatically assigned to your endpoint.

describedServer_arn :: Lens' DescribedServer Text

Specifies the unique Amazon Resource Name (ARN) of the server.