Network.AWS.Transfer.Types.EfsFileLocation

Description

Documentation

data EfsFileLocation

Reserved for future use.

See: newEfsFileLocation smart constructor.

Constructors

EfsFileLocation'

Fields

path :: Maybe Text

The pathname for the folder being used by a workflow.

fileSystemId :: Maybe Text

The ID of the file system, assigned by Amazon EFS.

newEfsFileLocation :: EfsFileLocation

Create a value of EfsFileLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:path:EfsFileLocation', efsFileLocation_path - The pathname for the folder being used by a workflow.

$sel:fileSystemId:EfsFileLocation', efsFileLocation_fileSystemId - The ID of the file system, assigned by Amazon EFS.

efsFileLocation_path :: Lens' EfsFileLocation (Maybe Text)

The pathname for the folder being used by a workflow.

efsFileLocation_fileSystemId :: Lens' EfsFileLocation (Maybe Text)

The ID of the file system, assigned by Amazon EFS.