Network.AWS.Transfer.Types.FileLocation

Description

Documentation

data FileLocation

Specifies the Amazon S3 or EFS file details to be used in the step.

See: newFileLocation smart constructor.

Constructors

FileLocation'

Fields

efsFileLocation :: Maybe EfsFileLocation

Specifies the Amazon EFS ID and the path for the file being used.

s3FileLocation :: Maybe S3FileLocation

Specifies the S3 details for the file being used, such as bucket, Etag, and so forth.

newFileLocation :: FileLocation

Create a value of FileLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:efsFileLocation:FileLocation', fileLocation_efsFileLocation - Specifies the Amazon EFS ID and the path for the file being used.

$sel:s3FileLocation:FileLocation', fileLocation_s3FileLocation - Specifies the S3 details for the file being used, such as bucket, Etag, and so forth.

fileLocation_efsFileLocation :: Lens' FileLocation (Maybe EfsFileLocation)

Specifies the Amazon EFS ID and the path for the file being used.

fileLocation_s3FileLocation :: Lens' FileLocation (Maybe S3FileLocation)

Specifies the S3 details for the file being used, such as bucket, Etag, and so forth.