Network.AWS.Transfer.Types.ListedServer

Description

Documentation

data ListedServer

Returns properties of a file transfer protocol-enabled server that was specified.

See: newListedServer smart constructor.

Constructors

ListedServer'

Fields

loggingRole :: Maybe Text

Specifies the Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that allows a server to turn on Amazon CloudWatch logging for Amazon S3 or Amazon EFS events. When set, user activity can be viewed in your CloudWatch logs.

state :: Maybe State

Specifies the condition of a server for the server that was described. A value of ONLINE indicates that the server can accept jobs and transfer files. A State value of OFFLINE means that the server cannot perform file transfer operations.

The states of STARTING and STOPPING indicate that the server is in an intermediate state, either not fully able to respond, or not fully offline. The values of START_FAILED or STOP_FAILED can indicate an error condition.

identityProviderType :: Maybe IdentityProviderType

Specifies the mode of authentication for a server. The default value is SERVICE_MANAGED, which allows you to store and access user credentials within the Amazon Web Services Transfer Family service.

Use AWS_DIRECTORY_SERVICE to provide access to Active Directory groups in Amazon Web Services Managed Active Directory or Microsoft Active Directory in your on-premises environment or in Amazon Web Services using AD Connectors. This option also requires you to provide a Directory ID using the IdentityProviderDetails parameter.

Use the API_GATEWAY value to integrate with an identity provider of your choosing. The API_GATEWAY setting requires you to provide an API Gateway endpoint URL to call for authentication using the IdentityProviderDetails parameter.

serverId :: Maybe Text

Specifies the unique system assigned identifier for the servers that were listed.

domain :: Maybe Domain

Specifies the domain of the storage system that is used for file transfers.

endpointType :: Maybe EndpointType

Specifies the type of VPC endpoint that your server is connected to. If your server is connected to a VPC endpoint, your server isn't accessible over the public internet.

userCount :: Maybe Int

Specifies the number of users that are assigned to a server you specified with the ServerId.

arn :: Text

Specifies the unique Amazon Resource Name (ARN) for a server to be listed.

newListedServer

Arguments

:: Text
$sel:arn:ListedServer'

-> ListedServer

Create a value of ListedServer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loggingRole:ListedServer', listedServer_loggingRole - Specifies the Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that allows a server to turn on Amazon CloudWatch logging for Amazon S3 or Amazon EFS events. When set, user activity can be viewed in your CloudWatch logs.

$sel:state:ListedServer', listedServer_state - Specifies the condition of a server for the server that was described. A value of ONLINE indicates that the server can accept jobs and transfer files. A State value of OFFLINE means that the server cannot perform file transfer operations.

The states of STARTING and STOPPING indicate that the server is in an intermediate state, either not fully able to respond, or not fully offline. The values of START_FAILED or STOP_FAILED can indicate an error condition.

$sel:identityProviderType:ListedServer', listedServer_identityProviderType - Specifies the mode of authentication for a server. The default value is SERVICE_MANAGED, which allows you to store and access user credentials within the Amazon Web Services Transfer Family service.

Use AWS_DIRECTORY_SERVICE to provide access to Active Directory groups in Amazon Web Services Managed Active Directory or Microsoft Active Directory in your on-premises environment or in Amazon Web Services using AD Connectors. This option also requires you to provide a Directory ID using the IdentityProviderDetails parameter.

Use the API_GATEWAY value to integrate with an identity provider of your choosing. The API_GATEWAY setting requires you to provide an API Gateway endpoint URL to call for authentication using the IdentityProviderDetails parameter.

$sel:serverId:ListedServer', listedServer_serverId - Specifies the unique system assigned identifier for the servers that were listed.

$sel:domain:ListedServer', listedServer_domain - Specifies the domain of the storage system that is used for file transfers.

$sel:endpointType:ListedServer', listedServer_endpointType - Specifies the type of VPC endpoint that your server is connected to. If your server is connected to a VPC endpoint, your server isn't accessible over the public internet.

$sel:userCount:ListedServer', listedServer_userCount - Specifies the number of users that are assigned to a server you specified with the ServerId.

$sel:arn:ListedServer', listedServer_arn - Specifies the unique Amazon Resource Name (ARN) for a server to be listed.

listedServer_loggingRole :: Lens' ListedServer (Maybe Text)

Specifies the Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that allows a server to turn on Amazon CloudWatch logging for Amazon S3 or Amazon EFS events. When set, user activity can be viewed in your CloudWatch logs.

listedServer_state :: Lens' ListedServer (Maybe State)

Specifies the condition of a server for the server that was described. A value of ONLINE indicates that the server can accept jobs and transfer files. A State value of OFFLINE means that the server cannot perform file transfer operations.

The states of STARTING and STOPPING indicate that the server is in an intermediate state, either not fully able to respond, or not fully offline. The values of START_FAILED or STOP_FAILED can indicate an error condition.

listedServer_identityProviderType :: Lens' ListedServer (Maybe IdentityProviderType)

Specifies the mode of authentication for a server. The default value is SERVICE_MANAGED, which allows you to store and access user credentials within the Amazon Web Services Transfer Family service.

Use AWS_DIRECTORY_SERVICE to provide access to Active Directory groups in Amazon Web Services Managed Active Directory or Microsoft Active Directory in your on-premises environment or in Amazon Web Services using AD Connectors. This option also requires you to provide a Directory ID using the IdentityProviderDetails parameter.

Use the API_GATEWAY value to integrate with an identity provider of your choosing. The API_GATEWAY setting requires you to provide an API Gateway endpoint URL to call for authentication using the IdentityProviderDetails parameter.

listedServer_serverId :: Lens' ListedServer (Maybe Text)

Specifies the unique system assigned identifier for the servers that were listed.

listedServer_domain :: Lens' ListedServer (Maybe Domain)

Specifies the domain of the storage system that is used for file transfers.

listedServer_endpointType :: Lens' ListedServer (Maybe EndpointType)

Specifies the type of VPC endpoint that your server is connected to. If your server is connected to a VPC endpoint, your server isn't accessible over the public internet.

listedServer_userCount :: Lens' ListedServer (Maybe Int)

Specifies the number of users that are assigned to a server you specified with the ServerId.

listedServer_arn :: Lens' ListedServer Text

Specifies the unique Amazon Resource Name (ARN) for a server to be listed.