Network.AWS.Transfer.Types.S3InputFileLocation

Description

Documentation

data S3InputFileLocation

Specifies the customer input S3 file location. If it is used inside copyStepDetails.DestinationFileLocation, it should be the S3 copy destination.

You need to provide the bucket and key. The key can represent either a path or a file. This is determined by whether or not you end the key value with the forward slash (/) character. If the final character is "/", then your file is copied to the folder, and its name does not change. If, rather, the final character is alphanumeric, your uploaded file is renamed to the path value. In this case, if a file with that name already exists, it is overwritten.

For example, if your path is shared-files/bob/, your uploaded files are copied to the shared-files/bob/, folder. If your path is shared-files/today, each uploaded file is copied to the shared-files folder and named today: each upload overwrites the previous version of the bob file.

See: newS3InputFileLocation smart constructor.

Constructors

S3InputFileLocation'

Fields

bucket :: Maybe Text

Specifies the S3 bucket for the customer input file.

key :: Maybe Text

The name assigned to the file when it was created in S3. You use the object key to retrieve the object.

newS3InputFileLocation :: S3InputFileLocation

Create a value of S3InputFileLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucket:S3InputFileLocation', s3InputFileLocation_bucket - Specifies the S3 bucket for the customer input file.

$sel:key:S3InputFileLocation', s3InputFileLocation_key - The name assigned to the file when it was created in S3. You use the object key to retrieve the object.

s3InputFileLocation_bucket :: Lens' S3InputFileLocation (Maybe Text)

Specifies the S3 bucket for the customer input file.

s3InputFileLocation_key :: Lens' S3InputFileLocation (Maybe Text)

The name assigned to the file when it was created in S3. You use the object key to retrieve the object.