Network.AWS.Transfer.Types.ServiceMetadata

Description

Documentation

data ServiceMetadata

A container object for the session details associated with a workflow.

See: newServiceMetadata smart constructor.

Constructors

ServiceMetadata'

Fields

userDetails :: UserDetails

The Server ID (ServerId), Session ID (SessionId) and user (UserName) make up the UserDetails.

newServiceMetadata

Arguments

:: UserDetails
$sel:userDetails:ServiceMetadata'

-> ServiceMetadata

Create a value of ServiceMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userDetails:ServiceMetadata', serviceMetadata_userDetails - The Server ID (ServerId), Session ID (SessionId) and user (UserName) make up the UserDetails.

serviceMetadata_userDetails :: Lens' ServiceMetadata UserDetails

The Server ID (ServerId), Session ID (SessionId) and user (UserName) make up the UserDetails.