Network.AWS.Transfer.Types.SshPublicKey

Description

Documentation

data SshPublicKey

Provides information about the public Secure Shell (SSH) key that is associated with a user account for the specific file transfer protocol-enabled server (as identified by ServerId). The information returned includes the date the key was imported, the public key contents, and the public key ID. A user can store more than one SSH public key associated with their user name on a specific server.

See: newSshPublicKey smart constructor.

Constructors

SshPublicKey'

Fields

dateImported :: POSIX

Specifies the date that the public key was added to the user account.

sshPublicKeyBody :: Text

Specifies the content of the SSH public key as specified by the PublicKeyId.

sshPublicKeyId :: Text

Specifies the SshPublicKeyId parameter contains the identifier of the public key.

newSshPublicKey

Arguments

:: UTCTime
$sel:dateImported:SshPublicKey'

-> Text
$sel:sshPublicKeyBody:SshPublicKey'

-> Text
$sel:sshPublicKeyId:SshPublicKey'

-> SshPublicKey

Create a value of SshPublicKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dateImported:SshPublicKey', sshPublicKey_dateImported - Specifies the date that the public key was added to the user account.

$sel:sshPublicKeyBody:SshPublicKey', sshPublicKey_sshPublicKeyBody - Specifies the content of the SSH public key as specified by the PublicKeyId.

$sel:sshPublicKeyId:SshPublicKey', sshPublicKey_sshPublicKeyId - Specifies the SshPublicKeyId parameter contains the identifier of the public key.

sshPublicKey_dateImported :: Lens' SshPublicKey UTCTime

Specifies the date that the public key was added to the user account.

sshPublicKey_sshPublicKeyBody :: Lens' SshPublicKey Text

Specifies the content of the SSH public key as specified by the PublicKeyId.

sshPublicKey_sshPublicKeyId :: Lens' SshPublicKey Text

Specifies the SshPublicKeyId parameter contains the identifier of the public key.