Network.AWS.Transfer.Types.WorkflowStep

Description

Documentation

data WorkflowStep

The basic building block of a workflow.

See: newWorkflowStep smart constructor.

Constructors

WorkflowStep'

Fields

tagStepDetails :: Maybe TagStepDetails

Details for a step that creates one or more tags.

You specify one or more tags: each tag contains a key/value pair.

deleteStepDetails :: Maybe DeleteStepDetails

Details for a step that deletes the file.

copyStepDetails :: Maybe CopyStepDetails

Details for a step that performs a file copy.

Consists of the following values:

A description

An S3 location for the destination of the file copy.

A flag that indicates whether or not to overwrite an existing file of the same name. The default is FALSE.

type' :: Maybe WorkflowStepType

Currently, the following step types are supported.

Copy: copy the file to another location

Custom: custom step with a lambda target

Delete: delete the file

Tag: add a tag to the file

customStepDetails :: Maybe CustomStepDetails

Details for a step that invokes a lambda function.

Consists of the lambda function name, target, and timeout (in seconds).

newWorkflowStep :: WorkflowStep

Create a value of WorkflowStep with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagStepDetails:WorkflowStep', workflowStep_tagStepDetails - Details for a step that creates one or more tags.

You specify one or more tags: each tag contains a key/value pair.

$sel:deleteStepDetails:WorkflowStep', workflowStep_deleteStepDetails - Details for a step that deletes the file.

$sel:copyStepDetails:WorkflowStep', workflowStep_copyStepDetails - Details for a step that performs a file copy.

Consists of the following values:

A description

An S3 location for the destination of the file copy.

A flag that indicates whether or not to overwrite an existing file of the same name. The default is FALSE.

$sel:type':WorkflowStep', workflowStep_type - Currently, the following step types are supported.

Copy: copy the file to another location

Custom: custom step with a lambda target

Delete: delete the file

Tag: add a tag to the file

$sel:customStepDetails:WorkflowStep', workflowStep_customStepDetails - Details for a step that invokes a lambda function.

Consists of the lambda function name, target, and timeout (in seconds).

workflowStep_tagStepDetails :: Lens' WorkflowStep (Maybe TagStepDetails)

Details for a step that creates one or more tags.

You specify one or more tags: each tag contains a key/value pair.

workflowStep_deleteStepDetails :: Lens' WorkflowStep (Maybe DeleteStepDetails)

Details for a step that deletes the file.

workflowStep_copyStepDetails :: Lens' WorkflowStep (Maybe CopyStepDetails)

Details for a step that performs a file copy.

Consists of the following values:

A description

An S3 location for the destination of the file copy.

A flag that indicates whether or not to overwrite an existing file of the same name. The default is FALSE.

workflowStep_type :: Lens' WorkflowStep (Maybe WorkflowStepType)

Currently, the following step types are supported.

Copy: copy the file to another location

Custom: custom step with a lambda target

Delete: delete the file

Tag: add a tag to the file

workflowStep_customStepDetails :: Lens' WorkflowStep (Maybe CustomStepDetails)

Details for a step that invokes a lambda function.

Consists of the lambda function name, target, and timeout (in seconds).