Network.AWS.Transfer.UpdateServer

Description

Updates the file transfer protocol-enabled server's properties after that server has been created.

The UpdateServer call returns the ServerId of the server you updated.

Creating a Request

data UpdateServer

See: newUpdateServer smart constructor.

Constructors

UpdateServer'

Fields

protocolDetails :: Maybe ProtocolDetails

The protocol settings that are configured for your server.

Use the PassiveIp parameter to indicate passive mode (for FTP and FTPS protocols). Enter a single dotted-quad IPv4 address, such as the external IP address of a firewall, router, or load balancer.

loggingRole :: Maybe Text

Specifies the Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that allows a server to turn on Amazon CloudWatch logging for Amazon S3 or Amazon EFS events. When set, user activity can be viewed in your CloudWatch logs.

hostKey :: Maybe (Sensitive Text)

The RSA private key as generated by ssh-keygen -N "" -m PEM -f my-new-server-key.

If you aren't planning to migrate existing users from an existing server to a new server, don't update the host key. Accidentally changing a server's host key can be disruptive.

For more information, see Change the host key for your SFTP-enabled server in the Amazon Web ServicesTransfer Family User Guide.

protocols :: Maybe (NonEmpty Protocol)

Specifies the file transfer protocol or protocols over which your file transfer protocol client can connect to your server's endpoint. The available protocols are:

Secure Shell (SSH) File Transfer Protocol (SFTP): File transfer over SSH

File Transfer Protocol Secure (FTPS): File transfer with TLS encryption

File Transfer Protocol (FTP): Unencrypted file transfer

If you select FTPS, you must choose a certificate stored in Amazon Web ServicesCertificate Manager (ACM) which will be used to identify your server when clients connect to it over FTPS.

If Protocol includes either FTP or FTPS, then the EndpointType must be VPC and the IdentityProviderType must be AWS_DIRECTORY_SERVICE or API_GATEWAY.

If Protocol includes FTP, then AddressAllocationIds cannot be associated.

If Protocol is set only to SFTP, the EndpointType can be set to PUBLIC and the IdentityProviderType can be set to SERVICE_MANAGED.

endpointType :: Maybe EndpointType

The type of endpoint that you want your server to use. You can choose to make your server's endpoint publicly accessible (PUBLIC) or host it inside your VPC. With an endpoint that is hosted in a VPC, you can restrict access to your server and resources only within your VPC or choose to make it internet facing by attaching Elastic IP addresses directly to it.

After May 19, 2021, you won't be able to create a server using EndpointType=VPC_ENDPOINT in your Amazon Web Servicesaccount if your account hasn't already done so before May 19, 2021. If you have already created servers with EndpointType=VPC_ENDPOINT in your Amazon Web Servicesaccount on or before May 19, 2021, you will not be affected. After this date, use EndpointType=VPC.

For more information, see https://docs.aws.amazon.com/transfer/latest/userguide/create-server-in-vpc.html#deprecate-vpc-endpoint.

It is recommended that you use VPC as the EndpointType. With this endpoint type, you have the option to directly associate up to three Elastic IPv4 addresses (BYO IP included) with your server's endpoint and use VPC security groups to restrict traffic by the client's public IP address. This is not possible with EndpointType set to VPC_ENDPOINT.

securityPolicyName :: Maybe Text

Specifies the name of the security policy that is attached to the server.

certificate :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon Web ServicesCertificate Manager (ACM) certificate. Required when Protocols is set to FTPS.

To request a new public certificate, see Request a public certificate in the Amazon Web ServicesCertificate Manager User Guide.

To import an existing certificate into ACM, see Importing certificates into ACM in the Amazon Web ServicesCertificate Manager User Guide.

To request a private certificate to use FTPS through private IP addresses, see Request a private certificate in the Amazon Web ServicesCertificate Manager User Guide.

Certificates with the following cryptographic algorithms and key sizes are supported:

2048-bit RSA (RSA_2048)

4096-bit RSA (RSA_4096)

Elliptic Prime Curve 256 bit (EC_prime256v1)

Elliptic Prime Curve 384 bit (EC_secp384r1)

Elliptic Prime Curve 521 bit (EC_secp521r1)

The certificate must be a valid SSL/TLS X.509 version 3 certificate with FQDN or IP address specified and information about the issuer.

identityProviderDetails :: Maybe IdentityProviderDetails

An array containing all of the information required to call a customer's authentication API method.

workflowDetails :: Maybe WorkflowDetails

Specifies the workflow ID for the workflow to assign and the execution role used for executing the workflow.

endpointDetails :: Maybe EndpointDetails

The virtual private cloud (VPC) endpoint settings that are configured for your server. When you host your endpoint within your VPC, you can make it accessible only to resources within your VPC, or you can attach Elastic IP addresses and make it accessible to clients over the internet. Your VPC's default security groups are automatically assigned to your endpoint.

serverId :: Text

A system-assigned unique identifier for a server instance that the user account is assigned to.

newUpdateServer

Arguments

:: Text
$sel:serverId:UpdateServer'

-> UpdateServer

Create a value of UpdateServer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocolDetails:UpdateServer', updateServer_protocolDetails - The protocol settings that are configured for your server.

Use the PassiveIp parameter to indicate passive mode (for FTP and FTPS protocols). Enter a single dotted-quad IPv4 address, such as the external IP address of a firewall, router, or load balancer.

$sel:loggingRole:UpdateServer', updateServer_loggingRole - Specifies the Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that allows a server to turn on Amazon CloudWatch logging for Amazon S3 or Amazon EFS events. When set, user activity can be viewed in your CloudWatch logs.

$sel:hostKey:UpdateServer', updateServer_hostKey - The RSA private key as generated by ssh-keygen -N "" -m PEM -f my-new-server-key.

If you aren't planning to migrate existing users from an existing server to a new server, don't update the host key. Accidentally changing a server's host key can be disruptive.

For more information, see Change the host key for your SFTP-enabled server in the Amazon Web ServicesTransfer Family User Guide.

$sel:protocols:UpdateServer', updateServer_protocols - Specifies the file transfer protocol or protocols over which your file transfer protocol client can connect to your server's endpoint. The available protocols are:

Secure Shell (SSH) File Transfer Protocol (SFTP): File transfer over SSH

File Transfer Protocol Secure (FTPS): File transfer with TLS encryption

File Transfer Protocol (FTP): Unencrypted file transfer

If you select FTPS, you must choose a certificate stored in Amazon Web ServicesCertificate Manager (ACM) which will be used to identify your server when clients connect to it over FTPS.

If Protocol includes either FTP or FTPS, then the EndpointType must be VPC and the IdentityProviderType must be AWS_DIRECTORY_SERVICE or API_GATEWAY.

If Protocol includes FTP, then AddressAllocationIds cannot be associated.

If Protocol is set only to SFTP, the EndpointType can be set to PUBLIC and the IdentityProviderType can be set to SERVICE_MANAGED.

$sel:endpointType:UpdateServer', updateServer_endpointType - The type of endpoint that you want your server to use. You can choose to make your server's endpoint publicly accessible (PUBLIC) or host it inside your VPC. With an endpoint that is hosted in a VPC, you can restrict access to your server and resources only within your VPC or choose to make it internet facing by attaching Elastic IP addresses directly to it.

After May 19, 2021, you won't be able to create a server using EndpointType=VPC_ENDPOINT in your Amazon Web Servicesaccount if your account hasn't already done so before May 19, 2021. If you have already created servers with EndpointType=VPC_ENDPOINT in your Amazon Web Servicesaccount on or before May 19, 2021, you will not be affected. After this date, use EndpointType=VPC.

For more information, see https://docs.aws.amazon.com/transfer/latest/userguide/create-server-in-vpc.html#deprecate-vpc-endpoint.

It is recommended that you use VPC as the EndpointType. With this endpoint type, you have the option to directly associate up to three Elastic IPv4 addresses (BYO IP included) with your server's endpoint and use VPC security groups to restrict traffic by the client's public IP address. This is not possible with EndpointType set to VPC_ENDPOINT.

$sel:securityPolicyName:UpdateServer', updateServer_securityPolicyName - Specifies the name of the security policy that is attached to the server.

$sel:certificate:UpdateServer', updateServer_certificate - The Amazon Resource Name (ARN) of the Amazon Web ServicesCertificate Manager (ACM) certificate. Required when Protocols is set to FTPS.

To request a new public certificate, see Request a public certificate in the Amazon Web ServicesCertificate Manager User Guide.

To import an existing certificate into ACM, see Importing certificates into ACM in the Amazon Web ServicesCertificate Manager User Guide.

To request a private certificate to use FTPS through private IP addresses, see Request a private certificate in the Amazon Web ServicesCertificate Manager User Guide.

Certificates with the following cryptographic algorithms and key sizes are supported:

2048-bit RSA (RSA_2048)

4096-bit RSA (RSA_4096)

Elliptic Prime Curve 256 bit (EC_prime256v1)

Elliptic Prime Curve 384 bit (EC_secp384r1)

Elliptic Prime Curve 521 bit (EC_secp521r1)

The certificate must be a valid SSL/TLS X.509 version 3 certificate with FQDN or IP address specified and information about the issuer.

$sel:identityProviderDetails:UpdateServer', updateServer_identityProviderDetails - An array containing all of the information required to call a customer's authentication API method.

$sel:workflowDetails:UpdateServer', updateServer_workflowDetails - Specifies the workflow ID for the workflow to assign and the execution role used for executing the workflow.

$sel:endpointDetails:UpdateServer', updateServer_endpointDetails - The virtual private cloud (VPC) endpoint settings that are configured for your server. When you host your endpoint within your VPC, you can make it accessible only to resources within your VPC, or you can attach Elastic IP addresses and make it accessible to clients over the internet. Your VPC's default security groups are automatically assigned to your endpoint.

$sel:serverId:UpdateServer', updateServer_serverId - A system-assigned unique identifier for a server instance that the user account is assigned to.

Request Lenses

updateServer_protocolDetails :: Lens' UpdateServer (Maybe ProtocolDetails)

The protocol settings that are configured for your server.

Use the PassiveIp parameter to indicate passive mode (for FTP and FTPS protocols). Enter a single dotted-quad IPv4 address, such as the external IP address of a firewall, router, or load balancer.

updateServer_loggingRole :: Lens' UpdateServer (Maybe Text)

Specifies the Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that allows a server to turn on Amazon CloudWatch logging for Amazon S3 or Amazon EFS events. When set, user activity can be viewed in your CloudWatch logs.

updateServer_hostKey :: Lens' UpdateServer (Maybe Text)

The RSA private key as generated by ssh-keygen -N "" -m PEM -f my-new-server-key.

If you aren't planning to migrate existing users from an existing server to a new server, don't update the host key. Accidentally changing a server's host key can be disruptive.

For more information, see Change the host key for your SFTP-enabled server in the Amazon Web ServicesTransfer Family User Guide.

updateServer_protocols :: Lens' UpdateServer (Maybe (NonEmpty Protocol))

Specifies the file transfer protocol or protocols over which your file transfer protocol client can connect to your server's endpoint. The available protocols are:

Secure Shell (SSH) File Transfer Protocol (SFTP): File transfer over SSH

File Transfer Protocol Secure (FTPS): File transfer with TLS encryption

File Transfer Protocol (FTP): Unencrypted file transfer

If you select FTPS, you must choose a certificate stored in Amazon Web ServicesCertificate Manager (ACM) which will be used to identify your server when clients connect to it over FTPS.

If Protocol includes either FTP or FTPS, then the EndpointType must be VPC and the IdentityProviderType must be AWS_DIRECTORY_SERVICE or API_GATEWAY.

If Protocol includes FTP, then AddressAllocationIds cannot be associated.

If Protocol is set only to SFTP, the EndpointType can be set to PUBLIC and the IdentityProviderType can be set to SERVICE_MANAGED.

updateServer_endpointType :: Lens' UpdateServer (Maybe EndpointType)

The type of endpoint that you want your server to use. You can choose to make your server's endpoint publicly accessible (PUBLIC) or host it inside your VPC. With an endpoint that is hosted in a VPC, you can restrict access to your server and resources only within your VPC or choose to make it internet facing by attaching Elastic IP addresses directly to it.

After May 19, 2021, you won't be able to create a server using EndpointType=VPC_ENDPOINT in your Amazon Web Servicesaccount if your account hasn't already done so before May 19, 2021. If you have already created servers with EndpointType=VPC_ENDPOINT in your Amazon Web Servicesaccount on or before May 19, 2021, you will not be affected. After this date, use EndpointType=VPC.

For more information, see https://docs.aws.amazon.com/transfer/latest/userguide/create-server-in-vpc.html#deprecate-vpc-endpoint.

It is recommended that you use VPC as the EndpointType. With this endpoint type, you have the option to directly associate up to three Elastic IPv4 addresses (BYO IP included) with your server's endpoint and use VPC security groups to restrict traffic by the client's public IP address. This is not possible with EndpointType set to VPC_ENDPOINT.

updateServer_securityPolicyName :: Lens' UpdateServer (Maybe Text)

Specifies the name of the security policy that is attached to the server.

updateServer_certificate :: Lens' UpdateServer (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Web ServicesCertificate Manager (ACM) certificate. Required when Protocols is set to FTPS.

To request a new public certificate, see Request a public certificate in the Amazon Web ServicesCertificate Manager User Guide.

To import an existing certificate into ACM, see Importing certificates into ACM in the Amazon Web ServicesCertificate Manager User Guide.

To request a private certificate to use FTPS through private IP addresses, see Request a private certificate in the Amazon Web ServicesCertificate Manager User Guide.

Certificates with the following cryptographic algorithms and key sizes are supported:

2048-bit RSA (RSA_2048)

4096-bit RSA (RSA_4096)

Elliptic Prime Curve 256 bit (EC_prime256v1)

Elliptic Prime Curve 384 bit (EC_secp384r1)

Elliptic Prime Curve 521 bit (EC_secp521r1)

The certificate must be a valid SSL/TLS X.509 version 3 certificate with FQDN or IP address specified and information about the issuer.

updateServer_identityProviderDetails :: Lens' UpdateServer (Maybe IdentityProviderDetails)

An array containing all of the information required to call a customer's authentication API method.

updateServer_workflowDetails :: Lens' UpdateServer (Maybe WorkflowDetails)

Specifies the workflow ID for the workflow to assign and the execution role used for executing the workflow.

updateServer_endpointDetails :: Lens' UpdateServer (Maybe EndpointDetails)

The virtual private cloud (VPC) endpoint settings that are configured for your server. When you host your endpoint within your VPC, you can make it accessible only to resources within your VPC, or you can attach Elastic IP addresses and make it accessible to clients over the internet. Your VPC's default security groups are automatically assigned to your endpoint.

updateServer_serverId :: Lens' UpdateServer Text

A system-assigned unique identifier for a server instance that the user account is assigned to.

Destructuring the Response

data UpdateServerResponse

See: newUpdateServerResponse smart constructor.

Constructors

UpdateServerResponse'

Fields

httpStatus :: Int

The response's http status code.

serverId :: Text

A system-assigned unique identifier for a server that the user account is assigned to.

newUpdateServerResponse

Arguments

:: Int
$sel:httpStatus:UpdateServerResponse'

-> Text
$sel:serverId:UpdateServer'

-> UpdateServerResponse

Create a value of UpdateServerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateServerResponse', updateServerResponse_httpStatus - The response's http status code.

$sel:serverId:UpdateServer', updateServerResponse_serverId - A system-assigned unique identifier for a server that the user account is assigned to.

Response Lenses

updateServerResponse_httpStatus :: Lens' UpdateServerResponse Int

The response's http status code.

updateServerResponse_serverId :: Lens' UpdateServerResponse Text

A system-assigned unique identifier for a server that the user account is assigned to.