Network.AWS.Translate

Description

Derived from API version 2017-07-01 of the AWS service descriptions, licensed under Apache 2.0.

Provides translation between one source language and another of the same set of languages.

Service Configuration

defaultService :: Service

API version 2017-07-01 of the Amazon Translate SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Translate.

InvalidRequestException

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request that you made is invalid. Check your request to determine why it's invalid and then retry the request.

UnsupportedLanguagePairException

_UnsupportedLanguagePairException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Translate does not support translation from the language of the source text into the requested target language. For more information, see how-to-error-msg.

DetectedLanguageLowConfidenceException

_DetectedLanguageLowConfidenceException :: AsError a => Getting (First ServiceError) a ServiceError

The confidence that Amazon Comprehend accurately detected the source language is low. If a low confidence level is acceptable for your application, you can use the language in the exception to call Amazon Translate again. For more information, see the DetectDominantLanguage operation in the Amazon Comprehend Developer Guide.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

There was a conflict processing the request. Try your request again.

InvalidParameterValueException

_InvalidParameterValueException :: AsError a => Getting (First ServiceError) a ServiceError

The value of the parameter is invalid. Review the value of the parameter you are using to correct it, and then retry your operation.

TooManyRequestsException

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

You have made too many requests within a short period of time. Wait for a short time and then try your request again.

ConcurrentModificationException

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

Another modification is being made. That modification must complete before you can make your change.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

An internal server error occurred. Retry your request.

ServiceUnavailableException

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The Amazon Translate service is temporarily unavailable. Please wait a bit and then retry your request.

InvalidFilterException

_InvalidFilterException :: AsError a => Getting (First ServiceError) a ServiceError

The filter specified for the operation is invalid. Specify a different filter.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resource you are looking for has not been found. Review the resource you're looking for and see if a different resource will accomplish your needs before retrying the revised request.

TextSizeLimitExceededException

_TextSizeLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The size of the text you submitted exceeds the size limit. Reduce the size of the text or use a smaller document and then retry your request.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The specified limit has been exceeded. Review your request and retry it with a quantity below the stated limit.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DescribeTextTranslationJob

data DescribeTextTranslationJob

See: newDescribeTextTranslationJob smart constructor.

Constructors

DescribeTextTranslationJob' Text

newDescribeTextTranslationJob

Arguments

:: Text
$sel:jobId:DescribeTextTranslationJob'

-> DescribeTextTranslationJob

Create a value of DescribeTextTranslationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:DescribeTextTranslationJob', describeTextTranslationJob_jobId - The identifier that Amazon Translate generated for the job. The StartTextTranslationJob operation returns this identifier in its response.

data DescribeTextTranslationJobResponse

See: newDescribeTextTranslationJobResponse smart constructor.

Constructors

DescribeTextTranslationJobResponse' (Maybe TextTranslationJobProperties) Int

newDescribeTextTranslationJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeTextTranslationJobResponse'

-> DescribeTextTranslationJobResponse

Create a value of DescribeTextTranslationJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:textTranslationJobProperties:DescribeTextTranslationJobResponse', describeTextTranslationJobResponse_textTranslationJobProperties - An object that contains the properties associated with an asynchronous batch translation job.

$sel:httpStatus:DescribeTextTranslationJobResponse', describeTextTranslationJobResponse_httpStatus - The response's http status code.

ListTerminologies (Paginated)

data ListTerminologies

See: newListTerminologies smart constructor.

Constructors

ListTerminologies' (Maybe Text) (Maybe Natural)

newListTerminologies :: ListTerminologies

Create a value of ListTerminologies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTerminologies', listTerminologies_nextToken - If the result of the request to ListTerminologies was truncated, include the NextToken to fetch the next group of custom terminologies.

$sel:maxResults:ListTerminologies', listTerminologies_maxResults - The maximum number of custom terminologies returned per list request.

data ListTerminologiesResponse

See: newListTerminologiesResponse smart constructor.

Constructors

ListTerminologiesResponse' (Maybe [TerminologyProperties]) (Maybe Text) Int

newListTerminologiesResponse

Arguments

:: Int
$sel:httpStatus:ListTerminologiesResponse'

-> ListTerminologiesResponse

Create a value of ListTerminologiesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:terminologyPropertiesList:ListTerminologiesResponse', listTerminologiesResponse_terminologyPropertiesList - The properties list of the custom terminologies returned on the list request.

$sel:nextToken:ListTerminologies', listTerminologiesResponse_nextToken - If the response to the ListTerminologies was truncated, the NextToken fetches the next group of custom terminologies.

$sel:httpStatus:ListTerminologiesResponse', listTerminologiesResponse_httpStatus - The response's http status code.

CreateParallelData

data CreateParallelData

See: newCreateParallelData smart constructor.

Constructors

CreateParallelData' (Maybe EncryptionKey) (Maybe Text) Text ParallelDataConfig Text

newCreateParallelData

Arguments

:: Text
$sel:name:CreateParallelData'

-> ParallelDataConfig
$sel:parallelDataConfig:CreateParallelData'

-> Text
$sel:clientToken:CreateParallelData'

-> CreateParallelData

Create a value of CreateParallelData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionKey:CreateParallelData', createParallelData_encryptionKey - Undocumented member.

$sel:description:CreateParallelData', createParallelData_description - A custom description for the parallel data resource in Amazon Translate.

$sel:name:CreateParallelData', createParallelData_name - A custom name for the parallel data resource in Amazon Translate. You must assign a name that is unique in the account and region.

$sel:parallelDataConfig:CreateParallelData', createParallelData_parallelDataConfig - Specifies the format and S3 location of the parallel data input file.

$sel:clientToken:CreateParallelData', createParallelData_clientToken - A unique identifier for the request. This token is automatically generated when you use Amazon Translate through an AWS SDK.

data CreateParallelDataResponse

See: newCreateParallelDataResponse smart constructor.

Constructors

CreateParallelDataResponse' (Maybe ParallelDataStatus) (Maybe Text) Int

newCreateParallelDataResponse

Arguments

:: Int
$sel:httpStatus:CreateParallelDataResponse'

-> CreateParallelDataResponse

Create a value of CreateParallelDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateParallelDataResponse', createParallelDataResponse_status - The status of the parallel data resource. When the resource is ready for you to use, the status is ACTIVE.

$sel:name:CreateParallelData', createParallelDataResponse_name - The custom name that you assigned to the parallel data resource.

$sel:httpStatus:CreateParallelDataResponse', createParallelDataResponse_httpStatus - The response's http status code.

UpdateParallelData

data UpdateParallelData

See: newUpdateParallelData smart constructor.

Constructors

UpdateParallelData' (Maybe Text) Text ParallelDataConfig Text

newUpdateParallelData

Arguments

:: Text
$sel:name:UpdateParallelData'

-> ParallelDataConfig
$sel:parallelDataConfig:UpdateParallelData'

-> Text
$sel:clientToken:UpdateParallelData'

-> UpdateParallelData

Create a value of UpdateParallelData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:UpdateParallelData', updateParallelData_description - A custom description for the parallel data resource in Amazon Translate.

$sel:name:UpdateParallelData', updateParallelData_name - The name of the parallel data resource being updated.

$sel:parallelDataConfig:UpdateParallelData', updateParallelData_parallelDataConfig - Specifies the format and S3 location of the parallel data input file.

$sel:clientToken:UpdateParallelData', updateParallelData_clientToken - A unique identifier for the request. This token is automatically generated when you use Amazon Translate through an AWS SDK.

data UpdateParallelDataResponse

See: newUpdateParallelDataResponse smart constructor.

Constructors

UpdateParallelDataResponse' (Maybe ParallelDataStatus) (Maybe Text) (Maybe POSIX) (Maybe ParallelDataStatus) Int

newUpdateParallelDataResponse

Arguments

:: Int
$sel:httpStatus:UpdateParallelDataResponse'

-> UpdateParallelDataResponse

Create a value of UpdateParallelDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UpdateParallelDataResponse', updateParallelDataResponse_status - The status of the parallel data resource that you are attempting to update. Your update request is accepted only if this status is either ACTIVE or FAILED.

$sel:name:UpdateParallelData', updateParallelDataResponse_name - The name of the parallel data resource being updated.

$sel:latestUpdateAttemptAt:UpdateParallelDataResponse', updateParallelDataResponse_latestUpdateAttemptAt - The time that the most recent update was attempted.

$sel:latestUpdateAttemptStatus:UpdateParallelDataResponse', updateParallelDataResponse_latestUpdateAttemptStatus - The status of the parallel data update attempt. When the updated parallel data resource is ready for you to use, the status is ACTIVE.

$sel:httpStatus:UpdateParallelDataResponse', updateParallelDataResponse_httpStatus - The response's http status code.

DeleteParallelData

data DeleteParallelData

See: newDeleteParallelData smart constructor.

Constructors

DeleteParallelData' Text

newDeleteParallelData

Arguments

:: Text
$sel:name:DeleteParallelData'

-> DeleteParallelData

Create a value of DeleteParallelData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DeleteParallelData', deleteParallelData_name - The name of the parallel data resource that is being deleted.

data DeleteParallelDataResponse

See: newDeleteParallelDataResponse smart constructor.

Constructors

DeleteParallelDataResponse' (Maybe ParallelDataStatus) (Maybe Text) Int

newDeleteParallelDataResponse

Arguments

:: Int
$sel:httpStatus:DeleteParallelDataResponse'

-> DeleteParallelDataResponse

Create a value of DeleteParallelDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DeleteParallelDataResponse', deleteParallelDataResponse_status - The status of the parallel data deletion.

$sel:name:DeleteParallelData', deleteParallelDataResponse_name - The name of the parallel data resource that is being deleted.

$sel:httpStatus:DeleteParallelDataResponse', deleteParallelDataResponse_httpStatus - The response's http status code.

GetParallelData

data GetParallelData

See: newGetParallelData smart constructor.

Constructors

GetParallelData' Text

newGetParallelData

Arguments

:: Text
$sel:name:GetParallelData'

-> GetParallelData

Create a value of GetParallelData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetParallelData', getParallelData_name - The name of the parallel data resource that is being retrieved.

data GetParallelDataResponse

See: newGetParallelDataResponse smart constructor.

Constructors

GetParallelDataResponse' (Maybe ParallelDataProperties) (Maybe ParallelDataDataLocation) (Maybe ParallelDataDataLocation) (Maybe ParallelDataDataLocation) Int

newGetParallelDataResponse

Arguments

:: Int
$sel:httpStatus:GetParallelDataResponse'

-> GetParallelDataResponse

Create a value of GetParallelDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parallelDataProperties:GetParallelDataResponse', getParallelDataResponse_parallelDataProperties - The properties of the parallel data resource that is being retrieved.

$sel:dataLocation:GetParallelDataResponse', getParallelDataResponse_dataLocation - The location of the most recent parallel data input file that was successfully imported into Amazon Translate. The location is returned as a presigned URL that has a 30 minute expiration.

$sel:auxiliaryDataLocation:GetParallelDataResponse', getParallelDataResponse_auxiliaryDataLocation - The Amazon S3 location of a file that provides any errors or warnings that were produced by your input file. This file was created when Amazon Translate attempted to create a parallel data resource. The location is returned as a presigned URL to that has a 30 minute expiration.

$sel:latestUpdateAttemptAuxiliaryDataLocation:GetParallelDataResponse', getParallelDataResponse_latestUpdateAttemptAuxiliaryDataLocation - The Amazon S3 location of a file that provides any errors or warnings that were produced by your input file. This file was created when Amazon Translate attempted to update a parallel data resource. The location is returned as a presigned URL to that has a 30 minute expiration.

$sel:httpStatus:GetParallelDataResponse', getParallelDataResponse_httpStatus - The response's http status code.

GetTerminology

data GetTerminology

See: newGetTerminology smart constructor.

Constructors

GetTerminology' Text TerminologyDataFormat

newGetTerminology

Arguments

:: Text
$sel:name:GetTerminology'

-> TerminologyDataFormat
$sel:terminologyDataFormat:GetTerminology'

-> GetTerminology

Create a value of GetTerminology with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetTerminology', getTerminology_name - The name of the custom terminology being retrieved.

$sel:terminologyDataFormat:GetTerminology', getTerminology_terminologyDataFormat - The data format of the custom terminology being retrieved, either CSV or TMX.

data GetTerminologyResponse

See: newGetTerminologyResponse smart constructor.

Constructors

GetTerminologyResponse' (Maybe TerminologyProperties) (Maybe TerminologyDataLocation) Int

newGetTerminologyResponse

Arguments

:: Int
$sel:httpStatus:GetTerminologyResponse'

-> GetTerminologyResponse

Create a value of GetTerminologyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:terminologyProperties:GetTerminologyResponse', getTerminologyResponse_terminologyProperties - The properties of the custom terminology being retrieved.

$sel:terminologyDataLocation:GetTerminologyResponse', getTerminologyResponse_terminologyDataLocation - The data location of the custom terminology being retrieved. The custom terminology file is returned in a presigned url that has a 30 minute expiration.

$sel:httpStatus:GetTerminologyResponse', getTerminologyResponse_httpStatus - The response's http status code.

TranslateText

data TranslateText

See: newTranslateText smart constructor.

Constructors

TranslateText' (Maybe [Text]) Text Text Text

newTranslateText

Arguments

:: Text
$sel:text:TranslateText'

-> Text
$sel:sourceLanguageCode:TranslateText'

-> Text
$sel:targetLanguageCode:TranslateText'

-> TranslateText

Create a value of TranslateText with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:terminologyNames:TranslateText', translateText_terminologyNames - The name of the terminology list file to be used in the TranslateText request. You can use 1 terminology list at most in a TranslateText request. Terminology lists can contain a maximum of 256 terms.

$sel:text:TranslateText', translateText_text - The text to translate. The text string can be a maximum of 5,000 bytes long. Depending on your character set, this may be fewer than 5,000 characters.

$sel:sourceLanguageCode:TranslateText', translateText_sourceLanguageCode - The language code for the language of the source text. The language must be a language supported by Amazon Translate. For a list of language codes, see what-is-languages.

To have Amazon Translate determine the source language of your text, you can specify auto in the SourceLanguageCode field. If you specify auto, Amazon Translate will call Amazon Comprehend to determine the source language.

$sel:targetLanguageCode:TranslateText', translateText_targetLanguageCode - The language code requested for the language of the target text. The language must be a language supported by Amazon Translate.

data TranslateTextResponse

See: newTranslateTextResponse smart constructor.

Constructors

TranslateTextResponse' (Maybe [AppliedTerminology]) Int Text Text Text

newTranslateTextResponse

Arguments

:: Int
$sel:httpStatus:TranslateTextResponse'

-> Text
$sel:translatedText:TranslateTextResponse'

-> Text
$sel:sourceLanguageCode:TranslateText'

-> Text
$sel:targetLanguageCode:TranslateText'

-> TranslateTextResponse

Create a value of TranslateTextResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appliedTerminologies:TranslateTextResponse', translateTextResponse_appliedTerminologies - The names of the custom terminologies applied to the input text by Amazon Translate for the translated text response.

$sel:httpStatus:TranslateTextResponse', translateTextResponse_httpStatus - The response's http status code.

$sel:translatedText:TranslateTextResponse', translateTextResponse_translatedText - The translated text.

$sel:sourceLanguageCode:TranslateText', translateTextResponse_sourceLanguageCode - The language code for the language of the source text.

$sel:targetLanguageCode:TranslateText', translateTextResponse_targetLanguageCode - The language code for the language of the target text.

ImportTerminology

data ImportTerminology

See: newImportTerminology smart constructor.

Constructors

ImportTerminology' (Maybe EncryptionKey) (Maybe Text) Text MergeStrategy TerminologyData

newImportTerminology

Arguments

:: Text
$sel:name:ImportTerminology'

-> MergeStrategy
$sel:mergeStrategy:ImportTerminology'

-> TerminologyData
$sel:terminologyData:ImportTerminology'

-> ImportTerminology

Create a value of ImportTerminology with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionKey:ImportTerminology', importTerminology_encryptionKey - The encryption key for the custom terminology being imported.

$sel:description:ImportTerminology', importTerminology_description - The description of the custom terminology being imported.

$sel:name:ImportTerminology', importTerminology_name - The name of the custom terminology being imported.

$sel:mergeStrategy:ImportTerminology', importTerminology_mergeStrategy - The merge strategy of the custom terminology being imported. Currently, only the OVERWRITE merge strategy is supported. In this case, the imported terminology will overwrite an existing terminology of the same name.

$sel:terminologyData:ImportTerminology', importTerminology_terminologyData - The terminology data for the custom terminology being imported.

data ImportTerminologyResponse

See: newImportTerminologyResponse smart constructor.

Constructors

ImportTerminologyResponse' (Maybe TerminologyProperties) Int

newImportTerminologyResponse

Arguments

:: Int
$sel:httpStatus:ImportTerminologyResponse'

-> ImportTerminologyResponse

Create a value of ImportTerminologyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:terminologyProperties:ImportTerminologyResponse', importTerminologyResponse_terminologyProperties - The properties of the custom terminology being imported.

$sel:httpStatus:ImportTerminologyResponse', importTerminologyResponse_httpStatus - The response's http status code.

StopTextTranslationJob

data StopTextTranslationJob

See: newStopTextTranslationJob smart constructor.

Constructors

StopTextTranslationJob' Text

newStopTextTranslationJob

Arguments

:: Text
$sel:jobId:StopTextTranslationJob'

-> StopTextTranslationJob

Create a value of StopTextTranslationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:StopTextTranslationJob', stopTextTranslationJob_jobId - The job ID of the job to be stopped.

data StopTextTranslationJobResponse

See: newStopTextTranslationJobResponse smart constructor.

Constructors

StopTextTranslationJobResponse' (Maybe Text) (Maybe JobStatus) Int

newStopTextTranslationJobResponse

Arguments

:: Int
$sel:httpStatus:StopTextTranslationJobResponse'

-> StopTextTranslationJobResponse

Create a value of StopTextTranslationJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:StopTextTranslationJob', stopTextTranslationJobResponse_jobId - The job ID of the stopped batch translation job.

$sel:jobStatus:StopTextTranslationJobResponse', stopTextTranslationJobResponse_jobStatus - The status of the designated job. Upon successful completion, the job's status will be STOPPED.

$sel:httpStatus:StopTextTranslationJobResponse', stopTextTranslationJobResponse_httpStatus - The response's http status code.

DeleteTerminology

data DeleteTerminology

See: newDeleteTerminology smart constructor.

Constructors

DeleteTerminology' Text

newDeleteTerminology

Arguments

:: Text
$sel:name:DeleteTerminology'

-> DeleteTerminology

Create a value of DeleteTerminology with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DeleteTerminology', deleteTerminology_name - The name of the custom terminology being deleted.

data DeleteTerminologyResponse

See: newDeleteTerminologyResponse smart constructor.

Constructors

DeleteTerminologyResponse'

newDeleteTerminologyResponse :: DeleteTerminologyResponse

Create a value of DeleteTerminologyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

ListTextTranslationJobs

data ListTextTranslationJobs

See: newListTextTranslationJobs smart constructor.

Constructors

ListTextTranslationJobs' (Maybe Text) (Maybe TextTranslationJobFilter) (Maybe Natural)

newListTextTranslationJobs :: ListTextTranslationJobs

Create a value of ListTextTranslationJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTextTranslationJobs', listTextTranslationJobs_nextToken - The token to request the next page of results.

$sel:filter':ListTextTranslationJobs', listTextTranslationJobs_filter - The parameters that specify which batch translation jobs to retrieve. Filters include job name, job status, and submission time. You can only set one filter at a time.

$sel:maxResults:ListTextTranslationJobs', listTextTranslationJobs_maxResults - The maximum number of results to return in each page. The default value is 100.

data ListTextTranslationJobsResponse

See: newListTextTranslationJobsResponse smart constructor.

Constructors

ListTextTranslationJobsResponse' (Maybe [TextTranslationJobProperties]) (Maybe Text) Int

newListTextTranslationJobsResponse

Arguments

:: Int
$sel:httpStatus:ListTextTranslationJobsResponse'

-> ListTextTranslationJobsResponse

Create a value of ListTextTranslationJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:textTranslationJobPropertiesList:ListTextTranslationJobsResponse', listTextTranslationJobsResponse_textTranslationJobPropertiesList - A list containing the properties of each job that is returned.

$sel:nextToken:ListTextTranslationJobs', listTextTranslationJobsResponse_nextToken - The token to use to retreive the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:ListTextTranslationJobsResponse', listTextTranslationJobsResponse_httpStatus - The response's http status code.

StartTextTranslationJob

data StartTextTranslationJob

See: newStartTextTranslationJob smart constructor.

Constructors

StartTextTranslationJob' (Maybe Text) (Maybe [Text]) (Maybe [Text]) InputDataConfig OutputDataConfig Text Text (NonEmpty Text) Text

newStartTextTranslationJob

Arguments

:: InputDataConfig
$sel:inputDataConfig:StartTextTranslationJob'

-> OutputDataConfig
$sel:outputDataConfig:StartTextTranslationJob'

-> Text
$sel:dataAccessRoleArn:StartTextTranslationJob'

-> Text
$sel:sourceLanguageCode:StartTextTranslationJob'

-> NonEmpty Text
$sel:targetLanguageCodes:StartTextTranslationJob'

-> Text
$sel:clientToken:StartTextTranslationJob'

-> StartTextTranslationJob

Create a value of StartTextTranslationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobName:StartTextTranslationJob', startTextTranslationJob_jobName - The name of the batch translation job to be performed.

$sel:parallelDataNames:StartTextTranslationJob', startTextTranslationJob_parallelDataNames - The names of the parallel data resources to use in the batch translation job. For a list of available parallel data resources, use the ListParallelData operation.

$sel:terminologyNames:StartTextTranslationJob', startTextTranslationJob_terminologyNames - The name of the terminology to use in the batch translation job. For a list of available terminologies, use the ListTerminologies operation.

$sel:inputDataConfig:StartTextTranslationJob', startTextTranslationJob_inputDataConfig - Specifies the format and S3 location of the input documents for the translation job.

$sel:outputDataConfig:StartTextTranslationJob', startTextTranslationJob_outputDataConfig - Specifies the S3 folder to which your job output will be saved.

$sel:dataAccessRoleArn:StartTextTranslationJob', startTextTranslationJob_dataAccessRoleArn - The Amazon Resource Name (ARN) of an AWS Identity Access and Management (IAM) role that grants Amazon Translate read access to your input data. For more nformation, see identity-and-access-management.

$sel:sourceLanguageCode:StartTextTranslationJob', startTextTranslationJob_sourceLanguageCode - The language code of the input language. For a list of language codes, see what-is-languages.

Amazon Translate does not automatically detect a source language during batch translation jobs.

$sel:targetLanguageCodes:StartTextTranslationJob', startTextTranslationJob_targetLanguageCodes - The language code of the output language.

$sel:clientToken:StartTextTranslationJob', startTextTranslationJob_clientToken - A unique identifier for the request. This token is auto-generated when using the Amazon Translate SDK.

data StartTextTranslationJobResponse

See: newStartTextTranslationJobResponse smart constructor.

Constructors

StartTextTranslationJobResponse' (Maybe Text) (Maybe JobStatus) Int

newStartTextTranslationJobResponse

Arguments

:: Int
$sel:httpStatus:StartTextTranslationJobResponse'

-> StartTextTranslationJobResponse

Create a value of StartTextTranslationJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:StartTextTranslationJobResponse', startTextTranslationJobResponse_jobId - The identifier generated for the job. To get the status of a job, use this ID with the DescribeTextTranslationJob operation.

$sel:jobStatus:StartTextTranslationJobResponse', startTextTranslationJobResponse_jobStatus - The status of the job. Possible values include:

SUBMITTED - The job has been received and is queued for processing.

IN_PROGRESS - Amazon Translate is processing the job.

COMPLETED - The job was successfully completed and the output is available.

COMPLETED_WITH_ERROR - The job was completed with errors. The errors can be analyzed in the job's output.

FAILED - The job did not complete. To get details, use the DescribeTextTranslationJob operation.

STOP_REQUESTED - The user who started the job has requested that it be stopped.

STOPPED - The job has been stopped.

$sel:httpStatus:StartTextTranslationJobResponse', startTextTranslationJobResponse_httpStatus - The response's http status code.

ListParallelData

data ListParallelData

See: newListParallelData smart constructor.

Constructors

ListParallelData' (Maybe Text) (Maybe Natural)

newListParallelData :: ListParallelData

Create a value of ListParallelData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListParallelData', listParallelData_nextToken - A string that specifies the next page of results to return in a paginated response.

$sel:maxResults:ListParallelData', listParallelData_maxResults - The maximum number of parallel data resources returned for each request.

data ListParallelDataResponse

See: newListParallelDataResponse smart constructor.

Constructors

ListParallelDataResponse' (Maybe [ParallelDataProperties]) (Maybe Text) Int

newListParallelDataResponse

Arguments

:: Int
$sel:httpStatus:ListParallelDataResponse'

-> ListParallelDataResponse

Create a value of ListParallelDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parallelDataPropertiesList:ListParallelDataResponse', listParallelDataResponse_parallelDataPropertiesList - The properties of the parallel data resources returned by this request.

$sel:nextToken:ListParallelData', listParallelDataResponse_nextToken - The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

$sel:httpStatus:ListParallelDataResponse', listParallelDataResponse_httpStatus - The response's http status code.

Types

EncryptionKeyType

newtype EncryptionKeyType

Constructors

EncryptionKeyType'

Fields

fromEncryptionKeyType :: Text

Bundled Patterns

pattern EncryptionKeyType_KMS :: EncryptionKeyType

JobStatus

newtype JobStatus

Constructors

JobStatus'

Fields

fromJobStatus :: Text

Bundled Patterns

pattern JobStatus_COMPLETED :: JobStatus

pattern JobStatus_COMPLETED_WITH_ERROR :: JobStatus

pattern JobStatus_FAILED :: JobStatus

pattern JobStatus_IN_PROGRESS :: JobStatus

pattern JobStatus_STOPPED :: JobStatus

pattern JobStatus_STOP_REQUESTED :: JobStatus

pattern JobStatus_SUBMITTED :: JobStatus

MergeStrategy

newtype MergeStrategy

Constructors

MergeStrategy'

Fields

fromMergeStrategy :: Text

Bundled Patterns

pattern MergeStrategy_OVERWRITE :: MergeStrategy

ParallelDataFormat

newtype ParallelDataFormat

Constructors

ParallelDataFormat'

Fields

fromParallelDataFormat :: Text

Bundled Patterns

pattern ParallelDataFormat_CSV :: ParallelDataFormat

pattern ParallelDataFormat_TMX :: ParallelDataFormat

pattern ParallelDataFormat_TSV :: ParallelDataFormat

ParallelDataStatus

newtype ParallelDataStatus

Constructors

ParallelDataStatus'

Fields

fromParallelDataStatus :: Text

Bundled Patterns

pattern ParallelDataStatus_ACTIVE :: ParallelDataStatus

pattern ParallelDataStatus_CREATING :: ParallelDataStatus

pattern ParallelDataStatus_DELETING :: ParallelDataStatus

pattern ParallelDataStatus_FAILED :: ParallelDataStatus

pattern ParallelDataStatus_UPDATING :: ParallelDataStatus

TerminologyDataFormat

newtype TerminologyDataFormat

Constructors

TerminologyDataFormat'

Fields

fromTerminologyDataFormat :: Text

Bundled Patterns

pattern TerminologyDataFormat_CSV :: TerminologyDataFormat

pattern TerminologyDataFormat_TMX :: TerminologyDataFormat

AppliedTerminology

data AppliedTerminology

The custom terminology applied to the input text by Amazon Translate for the translated text response. This is optional in the response and will only be present if you specified terminology input in the request. Currently, only one terminology can be applied per TranslateText request.

See: newAppliedTerminology smart constructor.

Constructors

AppliedTerminology' (Maybe [Term]) (Maybe Text)

newAppliedTerminology :: AppliedTerminology

Create a value of AppliedTerminology with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:terms:AppliedTerminology', appliedTerminology_terms - The specific terms of the custom terminology applied to the input text by Amazon Translate for the translated text response. A maximum of 250 terms will be returned, and the specific terms applied will be the first 250 terms in the source text.

$sel:name:AppliedTerminology', appliedTerminology_name - The name of the custom terminology applied to the input text by Amazon Translate for the translated text response.

EncryptionKey

data EncryptionKey

The encryption key used to encrypt this object.

See: newEncryptionKey smart constructor.

Constructors

EncryptionKey' EncryptionKeyType Text

newEncryptionKey

Arguments

:: EncryptionKeyType
$sel:type':EncryptionKey'

-> Text
$sel:id:EncryptionKey'

-> EncryptionKey

Create a value of EncryptionKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':EncryptionKey', encryptionKey_type - The type of encryption key used by Amazon Translate to encrypt custom terminologies.

$sel:id:EncryptionKey', encryptionKey_id - The Amazon Resource Name (ARN) of the encryption key being used to encrypt the custom terminology.

InputDataConfig

data InputDataConfig

The input configuration properties for requesting a batch translation job.

See: newInputDataConfig smart constructor.

Constructors

InputDataConfig' Text Text

newInputDataConfig

Arguments

:: Text
$sel:s3Uri:InputDataConfig'

-> Text
$sel:contentType:InputDataConfig'

-> InputDataConfig

Create a value of InputDataConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Uri:InputDataConfig', inputDataConfig_s3Uri - The URI of the AWS S3 folder that contains the input file. The folder must be in the same Region as the API endpoint you are calling.

$sel:contentType:InputDataConfig', inputDataConfig_contentType - Describes the format of the data that you submit to Amazon Translate as input. You can specify one of the following multipurpose internet mail extension (MIME) types:

text/html: The input data consists of one or more HTML files. Amazon Translate translates only the text that resides in the html element in each file.

text/plain: The input data consists of one or more unformatted text files. Amazon Translate translates every character in this type of input.

application/vnd.openxmlformats-officedocument.wordprocessingml.document: The input data consists of one or more Word documents (.docx).

application/vnd.openxmlformats-officedocument.presentationml.presentation: The input data consists of one or more PowerPoint Presentation files (.pptx).

application/vnd.openxmlformats-officedocument.spreadsheetml.sheet: The input data consists of one or more Excel Workbook files (.xlsx).

If you structure your input data as HTML, ensure that you set this parameter to text/html. By doing so, you cut costs by limiting the translation to the contents of the html element in each file. Otherwise, if you set this parameter to text/plain, your costs will cover the translation of every character.

JobDetails

data JobDetails

The number of documents successfully and unsuccessfully processed during a translation job.

See: newJobDetails smart constructor.

Constructors

JobDetails' (Maybe Int) (Maybe Int) (Maybe Int)

newJobDetails :: JobDetails

Create a value of JobDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:translatedDocumentsCount:JobDetails', jobDetails_translatedDocumentsCount - The number of documents successfully processed during a translation job.

$sel:documentsWithErrorsCount:JobDetails', jobDetails_documentsWithErrorsCount - The number of documents that could not be processed during a translation job.

$sel:inputDocumentsCount:JobDetails', jobDetails_inputDocumentsCount - The number of documents used as input in a translation job.

OutputDataConfig

data OutputDataConfig

The output configuration properties for a batch translation job.

See: newOutputDataConfig smart constructor.

Constructors

OutputDataConfig' Text

newOutputDataConfig

Arguments

:: Text
$sel:s3Uri:OutputDataConfig'

-> OutputDataConfig

Create a value of OutputDataConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Uri:OutputDataConfig', outputDataConfig_s3Uri - The URI of the S3 folder that contains a translation job's output file. The folder must be in the same Region as the API endpoint that you are calling.

ParallelDataConfig

data ParallelDataConfig

Specifies the format and S3 location of the parallel data input file.

See: newParallelDataConfig smart constructor.

Constructors

ParallelDataConfig' Text ParallelDataFormat

newParallelDataConfig

Arguments

:: Text
$sel:s3Uri:ParallelDataConfig'

-> ParallelDataFormat
$sel:format:ParallelDataConfig'

-> ParallelDataConfig

Create a value of ParallelDataConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Uri:ParallelDataConfig', parallelDataConfig_s3Uri - The URI of the Amazon S3 folder that contains the parallel data input file. The folder must be in the same Region as the API endpoint you are calling.

$sel:format:ParallelDataConfig', parallelDataConfig_format - The format of the parallel data input file.

ParallelDataDataLocation

data ParallelDataDataLocation

The location of the most recent parallel data input file that was successfully imported into Amazon Translate.

See: newParallelDataDataLocation smart constructor.

Constructors

ParallelDataDataLocation' Text Text

newParallelDataDataLocation

Arguments

:: Text
$sel:repositoryType:ParallelDataDataLocation'

-> Text
$sel:location:ParallelDataDataLocation'

-> ParallelDataDataLocation

Create a value of ParallelDataDataLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositoryType:ParallelDataDataLocation', parallelDataDataLocation_repositoryType - Describes the repository that contains the parallel data input file.

$sel:location:ParallelDataDataLocation', parallelDataDataLocation_location - The Amazon S3 location of the parallel data input file. The location is returned as a presigned URL to that has a 30 minute expiration.

ParallelDataProperties

data ParallelDataProperties

The properties of a parallel data resource.

See: newParallelDataProperties smart constructor.

Constructors

ParallelDataProperties' (Maybe ParallelDataStatus) (Maybe POSIX) (Maybe Integer) (Maybe Text) (Maybe [Text]) (Maybe POSIX) (Maybe Integer) (Maybe Integer) (Maybe Text) (Maybe Text) (Maybe POSIX) (Maybe EncryptionKey) (Maybe ParallelDataStatus) (Maybe Text) (Maybe Text) (Maybe Integer) (Maybe ParallelDataConfig)

newParallelDataProperties :: ParallelDataProperties

Create a value of ParallelDataProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ParallelDataProperties', parallelDataProperties_status - The status of the parallel data resource. When the parallel data is ready for you to use, the status is ACTIVE.

$sel:lastUpdatedAt:ParallelDataProperties', parallelDataProperties_lastUpdatedAt - The time at which the parallel data resource was last updated.

$sel:importedRecordCount:ParallelDataProperties', parallelDataProperties_importedRecordCount - The number of records successfully imported from the parallel data input file.

$sel:arn:ParallelDataProperties', parallelDataProperties_arn - The Amazon Resource Name (ARN) of the parallel data resource.

$sel:targetLanguageCodes:ParallelDataProperties', parallelDataProperties_targetLanguageCodes - The language codes for the target languages available in the parallel data file. All possible target languages are returned as an array.

$sel:createdAt:ParallelDataProperties', parallelDataProperties_createdAt - The time at which the parallel data resource was created.

$sel:failedRecordCount:ParallelDataProperties', parallelDataProperties_failedRecordCount - The number of records unsuccessfully imported from the parallel data input file.

$sel:importedDataSize:ParallelDataProperties', parallelDataProperties_importedDataSize - The number of UTF-8 characters that Amazon Translate imported from the parallel data input file. This number includes only the characters in your translation examples. It does not include characters that are used to format your file. For example, if you provided a Translation Memory Exchange (.tmx) file, this number does not include the tags.

$sel:name:ParallelDataProperties', parallelDataProperties_name - The custom name assigned to the parallel data resource.

$sel:sourceLanguageCode:ParallelDataProperties', parallelDataProperties_sourceLanguageCode - The source language of the translations in the parallel data file.

$sel:latestUpdateAttemptAt:ParallelDataProperties', parallelDataProperties_latestUpdateAttemptAt - The time that the most recent update was attempted.

$sel:encryptionKey:ParallelDataProperties', parallelDataProperties_encryptionKey - Undocumented member.

$sel:latestUpdateAttemptStatus:ParallelDataProperties', parallelDataProperties_latestUpdateAttemptStatus - The status of the most recent update attempt for the parallel data resource.

$sel:message:ParallelDataProperties', parallelDataProperties_message - Additional information from Amazon Translate about the parallel data resource.

$sel:description:ParallelDataProperties', parallelDataProperties_description - The description assigned to the parallel data resource.

$sel:skippedRecordCount:ParallelDataProperties', parallelDataProperties_skippedRecordCount - The number of items in the input file that Amazon Translate skipped when you created or updated the parallel data resource. For example, Amazon Translate skips empty records, empty target texts, and empty lines.

$sel:parallelDataConfig:ParallelDataProperties', parallelDataProperties_parallelDataConfig - Specifies the format and S3 location of the parallel data input file.

Term

data Term

The term being translated by the custom terminology.

See: newTerm smart constructor.

Constructors

Term' (Maybe Text) (Maybe Text)

newTerm :: Term

Create a value of Term with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetText:Term', term_targetText - The target text of the term being translated by the custom terminology.

$sel:sourceText:Term', term_sourceText - The source text of the term being translated by the custom terminology.

TerminologyData

data TerminologyData

The data associated with the custom terminology.

See: newTerminologyData smart constructor.

Constructors

TerminologyData' (Sensitive Base64) TerminologyDataFormat

newTerminologyData

Arguments

:: ByteString
$sel:file:TerminologyData'

-> TerminologyDataFormat
$sel:format:TerminologyData'

-> TerminologyData

Create a value of TerminologyData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:file:TerminologyData', terminologyData_file - The file containing the custom terminology data. Your version of the AWS SDK performs a Base64-encoding on this field before sending a request to the AWS service. Users of the SDK should not perform Base64-encoding themselves.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:format:TerminologyData', terminologyData_format - The data format of the custom terminology. Either CSV or TMX.

TerminologyDataLocation

data TerminologyDataLocation

The location of the custom terminology data.

See: newTerminologyDataLocation smart constructor.

Constructors

TerminologyDataLocation' Text Text

newTerminologyDataLocation

Arguments

:: Text
$sel:repositoryType:TerminologyDataLocation'

-> Text
$sel:location:TerminologyDataLocation'

-> TerminologyDataLocation

Create a value of TerminologyDataLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositoryType:TerminologyDataLocation', terminologyDataLocation_repositoryType - The repository type for the custom terminology data.

$sel:location:TerminologyDataLocation', terminologyDataLocation_location - The location of the custom terminology data.

TerminologyProperties

data TerminologyProperties

The properties of the custom terminology.

See: newTerminologyProperties smart constructor.

Constructors

TerminologyProperties' (Maybe Int) (Maybe POSIX) (Maybe Text) (Maybe [Text]) (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe Int) (Maybe EncryptionKey) (Maybe Text)

newTerminologyProperties :: TerminologyProperties

Create a value of TerminologyProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sizeBytes:TerminologyProperties', terminologyProperties_sizeBytes - The size of the file used when importing a custom terminology.

$sel:lastUpdatedAt:TerminologyProperties', terminologyProperties_lastUpdatedAt - The time at which the custom terminology was last update, based on the timestamp.

$sel:arn:TerminologyProperties', terminologyProperties_arn - The Amazon Resource Name (ARN) of the custom terminology.

$sel:targetLanguageCodes:TerminologyProperties', terminologyProperties_targetLanguageCodes - The language codes for the target languages available with the custom terminology file. All possible target languages are returned in array.

$sel:createdAt:TerminologyProperties', terminologyProperties_createdAt - The time at which the custom terminology was created, based on the timestamp.

$sel:name:TerminologyProperties', terminologyProperties_name - The name of the custom terminology.

$sel:sourceLanguageCode:TerminologyProperties', terminologyProperties_sourceLanguageCode - The language code for the source text of the translation request for which the custom terminology is being used.

$sel:termCount:TerminologyProperties', terminologyProperties_termCount - The number of terms included in the custom terminology.

$sel:encryptionKey:TerminologyProperties', terminologyProperties_encryptionKey - The encryption key for the custom terminology.

$sel:description:TerminologyProperties', terminologyProperties_description - The description of the custom terminology properties.

TextTranslationJobFilter

data TextTranslationJobFilter

Provides information for filtering a list of translation jobs. For more information, see ListTextTranslationJobs.

See: newTextTranslationJobFilter smart constructor.

Constructors

TextTranslationJobFilter' (Maybe POSIX) (Maybe POSIX) (Maybe Text) (Maybe JobStatus)

newTextTranslationJobFilter :: TextTranslationJobFilter

Create a value of TextTranslationJobFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:submittedBeforeTime:TextTranslationJobFilter', textTranslationJobFilter_submittedBeforeTime - Filters the list of jobs based on the time that the job was submitted for processing and returns only the jobs submitted before the specified time. Jobs are returned in ascending order, oldest to newest.

$sel:submittedAfterTime:TextTranslationJobFilter', textTranslationJobFilter_submittedAfterTime - Filters the list of jobs based on the time that the job was submitted for processing and returns only the jobs submitted after the specified time. Jobs are returned in descending order, newest to oldest.

$sel:jobName:TextTranslationJobFilter', textTranslationJobFilter_jobName - Filters the list of jobs by name.

$sel:jobStatus:TextTranslationJobFilter', textTranslationJobFilter_jobStatus - Filters the list of jobs based by job status.

TextTranslationJobProperties

data TextTranslationJobProperties

Provides information about a translation job.

See: newTextTranslationJobProperties smart constructor.

Constructors

TextTranslationJobProperties' (Maybe Text) (Maybe (NonEmpty Text)) (Maybe Text) (Maybe POSIX) (Maybe InputDataConfig) (Maybe [Text]) (Maybe [Text]) (Maybe Text) (Maybe POSIX) (Maybe OutputDataConfig) (Maybe JobDetails) (Maybe Text) (Maybe JobStatus) (Maybe Text)

newTextTranslationJobProperties :: TextTranslationJobProperties

Create a value of TextTranslationJobProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:TextTranslationJobProperties', textTranslationJobProperties_jobId - The ID of the translation job.

$sel:targetLanguageCodes:TextTranslationJobProperties', textTranslationJobProperties_targetLanguageCodes - The language code of the language of the target text. The language must be a language supported by Amazon Translate.

$sel:jobName:TextTranslationJobProperties', textTranslationJobProperties_jobName - The user-defined name of the translation job.

$sel:submittedTime:TextTranslationJobProperties', textTranslationJobProperties_submittedTime - The time at which the translation job was submitted.

$sel:inputDataConfig:TextTranslationJobProperties', textTranslationJobProperties_inputDataConfig - The input configuration properties that were specified when the job was requested.

$sel:parallelDataNames:TextTranslationJobProperties', textTranslationJobProperties_parallelDataNames - A list containing the names of the parallel data resources applied to the translation job.

$sel:terminologyNames:TextTranslationJobProperties', textTranslationJobProperties_terminologyNames - A list containing the names of the terminologies applied to a translation job. Only one terminology can be applied per StartTextTranslationJob request at this time.

$sel:sourceLanguageCode:TextTranslationJobProperties', textTranslationJobProperties_sourceLanguageCode - The language code of the language of the source text. The language must be a language supported by Amazon Translate.

$sel:endTime:TextTranslationJobProperties', textTranslationJobProperties_endTime - The time at which the translation job ended.

$sel:outputDataConfig:TextTranslationJobProperties', textTranslationJobProperties_outputDataConfig - The output configuration properties that were specified when the job was requested.

$sel:jobDetails:TextTranslationJobProperties', textTranslationJobProperties_jobDetails - The number of documents successfully and unsuccessfully processed during the translation job.

$sel:dataAccessRoleArn:TextTranslationJobProperties', textTranslationJobProperties_dataAccessRoleArn - The Amazon Resource Name (ARN) of an AWS Identity Access and Management (IAM) role that granted Amazon Translate read access to the job's input data.

$sel:jobStatus:TextTranslationJobProperties', textTranslationJobProperties_jobStatus - The status of the translation job.

$sel:message:TextTranslationJobProperties', textTranslationJobProperties_message - An explanation of any errors that may have occured during the translation job.