Network.AWS.Translate.CreateParallelData

Description

Creates a parallel data resource in Amazon Translate by importing an input file from Amazon S3. Parallel data files contain examples of source phrases and their translations from your translation memory. By adding parallel data, you can influence the style, tone, and word choice in your translation output.

Creating a Request

data CreateParallelData

See: newCreateParallelData smart constructor.

Constructors

CreateParallelData'

Fields

encryptionKey :: Maybe EncryptionKey

description :: Maybe Text

A custom description for the parallel data resource in Amazon Translate.

name :: Text

A custom name for the parallel data resource in Amazon Translate. You must assign a name that is unique in the account and region.

parallelDataConfig :: ParallelDataConfig

Specifies the format and S3 location of the parallel data input file.

clientToken :: Text

A unique identifier for the request. This token is automatically generated when you use Amazon Translate through an AWS SDK.

newCreateParallelData

Arguments

:: Text
$sel:name:CreateParallelData'

-> ParallelDataConfig
$sel:parallelDataConfig:CreateParallelData'

-> Text
$sel:clientToken:CreateParallelData'

-> CreateParallelData

Create a value of CreateParallelData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionKey:CreateParallelData', createParallelData_encryptionKey - Undocumented member.

$sel:description:CreateParallelData', createParallelData_description - A custom description for the parallel data resource in Amazon Translate.

$sel:name:CreateParallelData', createParallelData_name - A custom name for the parallel data resource in Amazon Translate. You must assign a name that is unique in the account and region.

$sel:parallelDataConfig:CreateParallelData', createParallelData_parallelDataConfig - Specifies the format and S3 location of the parallel data input file.

$sel:clientToken:CreateParallelData', createParallelData_clientToken - A unique identifier for the request. This token is automatically generated when you use Amazon Translate through an AWS SDK.

Request Lenses

createParallelData_encryptionKey :: Lens' CreateParallelData (Maybe EncryptionKey)

Undocumented member.

createParallelData_description :: Lens' CreateParallelData (Maybe Text)

A custom description for the parallel data resource in Amazon Translate.

createParallelData_name :: Lens' CreateParallelData Text

A custom name for the parallel data resource in Amazon Translate. You must assign a name that is unique in the account and region.

createParallelData_parallelDataConfig :: Lens' CreateParallelData ParallelDataConfig

Specifies the format and S3 location of the parallel data input file.

createParallelData_clientToken :: Lens' CreateParallelData Text

A unique identifier for the request. This token is automatically generated when you use Amazon Translate through an AWS SDK.

Destructuring the Response

data CreateParallelDataResponse

See: newCreateParallelDataResponse smart constructor.

Constructors

CreateParallelDataResponse'

Fields

status :: Maybe ParallelDataStatus

The status of the parallel data resource. When the resource is ready for you to use, the status is ACTIVE.

name :: Maybe Text

The custom name that you assigned to the parallel data resource.

httpStatus :: Int

The response's http status code.

newCreateParallelDataResponse

Arguments

:: Int
$sel:httpStatus:CreateParallelDataResponse'

-> CreateParallelDataResponse

Create a value of CreateParallelDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateParallelDataResponse', createParallelDataResponse_status - The status of the parallel data resource. When the resource is ready for you to use, the status is ACTIVE.

$sel:name:CreateParallelData', createParallelDataResponse_name - The custom name that you assigned to the parallel data resource.

$sel:httpStatus:CreateParallelDataResponse', createParallelDataResponse_httpStatus - The response's http status code.

Response Lenses

createParallelDataResponse_status :: Lens' CreateParallelDataResponse (Maybe ParallelDataStatus)

The status of the parallel data resource. When the resource is ready for you to use, the status is ACTIVE.

createParallelDataResponse_name :: Lens' CreateParallelDataResponse (Maybe Text)

The custom name that you assigned to the parallel data resource.

createParallelDataResponse_httpStatus :: Lens' CreateParallelDataResponse Int

The response's http status code.