Network.AWS.Translate.GetParallelData

Description

Provides information about a parallel data resource.

Creating a Request

data GetParallelData

See: newGetParallelData smart constructor.

Constructors

GetParallelData'

Fields

name :: Text

The name of the parallel data resource that is being retrieved.

newGetParallelData

Arguments

:: Text
$sel:name:GetParallelData'

-> GetParallelData

Create a value of GetParallelData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetParallelData', getParallelData_name - The name of the parallel data resource that is being retrieved.

Request Lenses

getParallelData_name :: Lens' GetParallelData Text

The name of the parallel data resource that is being retrieved.

Destructuring the Response

data GetParallelDataResponse

See: newGetParallelDataResponse smart constructor.

Constructors

GetParallelDataResponse'

Fields

parallelDataProperties :: Maybe ParallelDataProperties

The properties of the parallel data resource that is being retrieved.

dataLocation :: Maybe ParallelDataDataLocation

The location of the most recent parallel data input file that was successfully imported into Amazon Translate. The location is returned as a presigned URL that has a 30 minute expiration.

auxiliaryDataLocation :: Maybe ParallelDataDataLocation

The Amazon S3 location of a file that provides any errors or warnings that were produced by your input file. This file was created when Amazon Translate attempted to create a parallel data resource. The location is returned as a presigned URL to that has a 30 minute expiration.

latestUpdateAttemptAuxiliaryDataLocation :: Maybe ParallelDataDataLocation

The Amazon S3 location of a file that provides any errors or warnings that were produced by your input file. This file was created when Amazon Translate attempted to update a parallel data resource. The location is returned as a presigned URL to that has a 30 minute expiration.

httpStatus :: Int

The response's http status code.

newGetParallelDataResponse

Arguments

:: Int
$sel:httpStatus:GetParallelDataResponse'

-> GetParallelDataResponse

Create a value of GetParallelDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parallelDataProperties:GetParallelDataResponse', getParallelDataResponse_parallelDataProperties - The properties of the parallel data resource that is being retrieved.

$sel:dataLocation:GetParallelDataResponse', getParallelDataResponse_dataLocation - The location of the most recent parallel data input file that was successfully imported into Amazon Translate. The location is returned as a presigned URL that has a 30 minute expiration.

$sel:auxiliaryDataLocation:GetParallelDataResponse', getParallelDataResponse_auxiliaryDataLocation - The Amazon S3 location of a file that provides any errors or warnings that were produced by your input file. This file was created when Amazon Translate attempted to create a parallel data resource. The location is returned as a presigned URL to that has a 30 minute expiration.

$sel:latestUpdateAttemptAuxiliaryDataLocation:GetParallelDataResponse', getParallelDataResponse_latestUpdateAttemptAuxiliaryDataLocation - The Amazon S3 location of a file that provides any errors or warnings that were produced by your input file. This file was created when Amazon Translate attempted to update a parallel data resource. The location is returned as a presigned URL to that has a 30 minute expiration.

$sel:httpStatus:GetParallelDataResponse', getParallelDataResponse_httpStatus - The response's http status code.

Response Lenses

getParallelDataResponse_parallelDataProperties :: Lens' GetParallelDataResponse (Maybe ParallelDataProperties)

The properties of the parallel data resource that is being retrieved.

getParallelDataResponse_dataLocation :: Lens' GetParallelDataResponse (Maybe ParallelDataDataLocation)

The location of the most recent parallel data input file that was successfully imported into Amazon Translate. The location is returned as a presigned URL that has a 30 minute expiration.

getParallelDataResponse_auxiliaryDataLocation :: Lens' GetParallelDataResponse (Maybe ParallelDataDataLocation)

The Amazon S3 location of a file that provides any errors or warnings that were produced by your input file. This file was created when Amazon Translate attempted to create a parallel data resource. The location is returned as a presigned URL to that has a 30 minute expiration.

getParallelDataResponse_latestUpdateAttemptAuxiliaryDataLocation :: Lens' GetParallelDataResponse (Maybe ParallelDataDataLocation)

The Amazon S3 location of a file that provides any errors or warnings that were produced by your input file. This file was created when Amazon Translate attempted to update a parallel data resource. The location is returned as a presigned URL to that has a 30 minute expiration.

getParallelDataResponse_httpStatus :: Lens' GetParallelDataResponse Int

The response's http status code.