Network.AWS.Translate.GetTerminology

Description

Retrieves a custom terminology.

Creating a Request

data GetTerminology

See: newGetTerminology smart constructor.

Constructors

GetTerminology'

Fields

name :: Text

The name of the custom terminology being retrieved.

terminologyDataFormat :: TerminologyDataFormat

The data format of the custom terminology being retrieved, either CSV or TMX.

newGetTerminology

Arguments

:: Text
$sel:name:GetTerminology'

-> TerminologyDataFormat
$sel:terminologyDataFormat:GetTerminology'

-> GetTerminology

Create a value of GetTerminology with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetTerminology', getTerminology_name - The name of the custom terminology being retrieved.

$sel:terminologyDataFormat:GetTerminology', getTerminology_terminologyDataFormat - The data format of the custom terminology being retrieved, either CSV or TMX.

Request Lenses

getTerminology_name :: Lens' GetTerminology Text

The name of the custom terminology being retrieved.

getTerminology_terminologyDataFormat :: Lens' GetTerminology TerminologyDataFormat

The data format of the custom terminology being retrieved, either CSV or TMX.

Destructuring the Response

data GetTerminologyResponse

See: newGetTerminologyResponse smart constructor.

Constructors

GetTerminologyResponse'

Fields

terminologyDataLocation :: Maybe TerminologyDataLocation

The data location of the custom terminology being retrieved. The custom terminology file is returned in a presigned url that has a 30 minute expiration.

terminologyProperties :: Maybe TerminologyProperties

The properties of the custom terminology being retrieved.

httpStatus :: Int

The response's http status code.

newGetTerminologyResponse

Arguments

:: Int
$sel:httpStatus:GetTerminologyResponse'

-> GetTerminologyResponse

Create a value of GetTerminologyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:terminologyDataLocation:GetTerminologyResponse', getTerminologyResponse_terminologyDataLocation - The data location of the custom terminology being retrieved. The custom terminology file is returned in a presigned url that has a 30 minute expiration.

$sel:terminologyProperties:GetTerminologyResponse', getTerminologyResponse_terminologyProperties - The properties of the custom terminology being retrieved.

$sel:httpStatus:GetTerminologyResponse', getTerminologyResponse_httpStatus - The response's http status code.

Response Lenses

getTerminologyResponse_terminologyDataLocation :: Lens' GetTerminologyResponse (Maybe TerminologyDataLocation)

The data location of the custom terminology being retrieved. The custom terminology file is returned in a presigned url that has a 30 minute expiration.

getTerminologyResponse_terminologyProperties :: Lens' GetTerminologyResponse (Maybe TerminologyProperties)

The properties of the custom terminology being retrieved.

getTerminologyResponse_httpStatus :: Lens' GetTerminologyResponse Int

The response's http status code.