Network.AWS.Translate.Lens

Description

Operations

DescribeTextTranslationJob

describeTextTranslationJob_jobId :: Lens' DescribeTextTranslationJob Text

The identifier that Amazon Translate generated for the job. The StartTextTranslationJob operation returns this identifier in its response.

describeTextTranslationJobResponse_textTranslationJobProperties :: Lens' DescribeTextTranslationJobResponse (Maybe TextTranslationJobProperties)

An object that contains the properties associated with an asynchronous batch translation job.

describeTextTranslationJobResponse_httpStatus :: Lens' DescribeTextTranslationJobResponse Int

The response's http status code.

ListTerminologies

listTerminologies_nextToken :: Lens' ListTerminologies (Maybe Text)

If the result of the request to ListTerminologies was truncated, include the NextToken to fetch the next group of custom terminologies.

listTerminologies_maxResults :: Lens' ListTerminologies (Maybe Natural)

The maximum number of custom terminologies returned per list request.

listTerminologiesResponse_terminologyPropertiesList :: Lens' ListTerminologiesResponse (Maybe [TerminologyProperties])

The properties list of the custom terminologies returned on the list request.

listTerminologiesResponse_nextToken :: Lens' ListTerminologiesResponse (Maybe Text)

If the response to the ListTerminologies was truncated, the NextToken fetches the next group of custom terminologies.

listTerminologiesResponse_httpStatus :: Lens' ListTerminologiesResponse Int

The response's http status code.

CreateParallelData

createParallelData_encryptionKey :: Lens' CreateParallelData (Maybe EncryptionKey)

Undocumented member.

createParallelData_description :: Lens' CreateParallelData (Maybe Text)

A custom description for the parallel data resource in Amazon Translate.

createParallelData_name :: Lens' CreateParallelData Text

A custom name for the parallel data resource in Amazon Translate. You must assign a name that is unique in the account and region.

createParallelData_parallelDataConfig :: Lens' CreateParallelData ParallelDataConfig

Specifies the format and S3 location of the parallel data input file.

createParallelData_clientToken :: Lens' CreateParallelData Text

A unique identifier for the request. This token is automatically generated when you use Amazon Translate through an AWS SDK.

createParallelDataResponse_status :: Lens' CreateParallelDataResponse (Maybe ParallelDataStatus)

The status of the parallel data resource. When the resource is ready for you to use, the status is ACTIVE.

createParallelDataResponse_name :: Lens' CreateParallelDataResponse (Maybe Text)

The custom name that you assigned to the parallel data resource.

createParallelDataResponse_httpStatus :: Lens' CreateParallelDataResponse Int

The response's http status code.

UpdateParallelData

updateParallelData_description :: Lens' UpdateParallelData (Maybe Text)

A custom description for the parallel data resource in Amazon Translate.

updateParallelData_name :: Lens' UpdateParallelData Text

The name of the parallel data resource being updated.

updateParallelData_parallelDataConfig :: Lens' UpdateParallelData ParallelDataConfig

Specifies the format and S3 location of the parallel data input file.

updateParallelData_clientToken :: Lens' UpdateParallelData Text

A unique identifier for the request. This token is automatically generated when you use Amazon Translate through an AWS SDK.

updateParallelDataResponse_status :: Lens' UpdateParallelDataResponse (Maybe ParallelDataStatus)

The status of the parallel data resource that you are attempting to update. Your update request is accepted only if this status is either ACTIVE or FAILED.

updateParallelDataResponse_name :: Lens' UpdateParallelDataResponse (Maybe Text)

The name of the parallel data resource being updated.

updateParallelDataResponse_latestUpdateAttemptAt :: Lens' UpdateParallelDataResponse (Maybe UTCTime)

The time that the most recent update was attempted.

updateParallelDataResponse_latestUpdateAttemptStatus :: Lens' UpdateParallelDataResponse (Maybe ParallelDataStatus)

The status of the parallel data update attempt. When the updated parallel data resource is ready for you to use, the status is ACTIVE.

updateParallelDataResponse_httpStatus :: Lens' UpdateParallelDataResponse Int

The response's http status code.

DeleteParallelData

deleteParallelData_name :: Lens' DeleteParallelData Text

The name of the parallel data resource that is being deleted.

deleteParallelDataResponse_status :: Lens' DeleteParallelDataResponse (Maybe ParallelDataStatus)

The status of the parallel data deletion.

deleteParallelDataResponse_name :: Lens' DeleteParallelDataResponse (Maybe Text)

The name of the parallel data resource that is being deleted.

deleteParallelDataResponse_httpStatus :: Lens' DeleteParallelDataResponse Int

The response's http status code.

GetParallelData

getParallelData_name :: Lens' GetParallelData Text

The name of the parallel data resource that is being retrieved.

getParallelDataResponse_parallelDataProperties :: Lens' GetParallelDataResponse (Maybe ParallelDataProperties)

The properties of the parallel data resource that is being retrieved.

getParallelDataResponse_dataLocation :: Lens' GetParallelDataResponse (Maybe ParallelDataDataLocation)

The location of the most recent parallel data input file that was successfully imported into Amazon Translate. The location is returned as a presigned URL that has a 30 minute expiration.

getParallelDataResponse_auxiliaryDataLocation :: Lens' GetParallelDataResponse (Maybe ParallelDataDataLocation)

The Amazon S3 location of a file that provides any errors or warnings that were produced by your input file. This file was created when Amazon Translate attempted to create a parallel data resource. The location is returned as a presigned URL to that has a 30 minute expiration.

getParallelDataResponse_latestUpdateAttemptAuxiliaryDataLocation :: Lens' GetParallelDataResponse (Maybe ParallelDataDataLocation)

The Amazon S3 location of a file that provides any errors or warnings that were produced by your input file. This file was created when Amazon Translate attempted to update a parallel data resource. The location is returned as a presigned URL to that has a 30 minute expiration.

getParallelDataResponse_httpStatus :: Lens' GetParallelDataResponse Int

The response's http status code.

GetTerminology

getTerminology_name :: Lens' GetTerminology Text

The name of the custom terminology being retrieved.

getTerminology_terminologyDataFormat :: Lens' GetTerminology TerminologyDataFormat

The data format of the custom terminology being retrieved, either CSV or TMX.

getTerminologyResponse_terminologyProperties :: Lens' GetTerminologyResponse (Maybe TerminologyProperties)

The properties of the custom terminology being retrieved.

getTerminologyResponse_terminologyDataLocation :: Lens' GetTerminologyResponse (Maybe TerminologyDataLocation)

The data location of the custom terminology being retrieved. The custom terminology file is returned in a presigned url that has a 30 minute expiration.

getTerminologyResponse_httpStatus :: Lens' GetTerminologyResponse Int

The response's http status code.

TranslateText

translateText_terminologyNames :: Lens' TranslateText (Maybe [Text])

The name of the terminology list file to be used in the TranslateText request. You can use 1 terminology list at most in a TranslateText request. Terminology lists can contain a maximum of 256 terms.

translateText_text :: Lens' TranslateText Text

The text to translate. The text string can be a maximum of 5,000 bytes long. Depending on your character set, this may be fewer than 5,000 characters.

translateText_sourceLanguageCode :: Lens' TranslateText Text

The language code for the language of the source text. The language must be a language supported by Amazon Translate. For a list of language codes, see what-is-languages.

To have Amazon Translate determine the source language of your text, you can specify auto in the SourceLanguageCode field. If you specify auto, Amazon Translate will call Amazon Comprehend to determine the source language.

translateText_targetLanguageCode :: Lens' TranslateText Text

The language code requested for the language of the target text. The language must be a language supported by Amazon Translate.

translateTextResponse_appliedTerminologies :: Lens' TranslateTextResponse (Maybe [AppliedTerminology])

The names of the custom terminologies applied to the input text by Amazon Translate for the translated text response.

translateTextResponse_httpStatus :: Lens' TranslateTextResponse Int

The response's http status code.

translateTextResponse_translatedText :: Lens' TranslateTextResponse Text

The translated text.

translateTextResponse_sourceLanguageCode :: Lens' TranslateTextResponse Text

The language code for the language of the source text.

translateTextResponse_targetLanguageCode :: Lens' TranslateTextResponse Text

The language code for the language of the target text.

ImportTerminology

importTerminology_encryptionKey :: Lens' ImportTerminology (Maybe EncryptionKey)

The encryption key for the custom terminology being imported.

importTerminology_description :: Lens' ImportTerminology (Maybe Text)

The description of the custom terminology being imported.

importTerminology_name :: Lens' ImportTerminology Text

The name of the custom terminology being imported.

importTerminology_mergeStrategy :: Lens' ImportTerminology MergeStrategy

The merge strategy of the custom terminology being imported. Currently, only the OVERWRITE merge strategy is supported. In this case, the imported terminology will overwrite an existing terminology of the same name.

importTerminology_terminologyData :: Lens' ImportTerminology TerminologyData

The terminology data for the custom terminology being imported.

importTerminologyResponse_terminologyProperties :: Lens' ImportTerminologyResponse (Maybe TerminologyProperties)

The properties of the custom terminology being imported.

importTerminologyResponse_httpStatus :: Lens' ImportTerminologyResponse Int

The response's http status code.

StopTextTranslationJob

stopTextTranslationJob_jobId :: Lens' StopTextTranslationJob Text

The job ID of the job to be stopped.

stopTextTranslationJobResponse_jobId :: Lens' StopTextTranslationJobResponse (Maybe Text)

The job ID of the stopped batch translation job.

stopTextTranslationJobResponse_jobStatus :: Lens' StopTextTranslationJobResponse (Maybe JobStatus)

The status of the designated job. Upon successful completion, the job's status will be STOPPED.

stopTextTranslationJobResponse_httpStatus :: Lens' StopTextTranslationJobResponse Int

The response's http status code.

DeleteTerminology

deleteTerminology_name :: Lens' DeleteTerminology Text

The name of the custom terminology being deleted.

ListTextTranslationJobs

listTextTranslationJobs_nextToken :: Lens' ListTextTranslationJobs (Maybe Text)

The token to request the next page of results.

listTextTranslationJobs_filter :: Lens' ListTextTranslationJobs (Maybe TextTranslationJobFilter)

The parameters that specify which batch translation jobs to retrieve. Filters include job name, job status, and submission time. You can only set one filter at a time.

listTextTranslationJobs_maxResults :: Lens' ListTextTranslationJobs (Maybe Natural)

The maximum number of results to return in each page. The default value is 100.

listTextTranslationJobsResponse_textTranslationJobPropertiesList :: Lens' ListTextTranslationJobsResponse (Maybe [TextTranslationJobProperties])

A list containing the properties of each job that is returned.

listTextTranslationJobsResponse_nextToken :: Lens' ListTextTranslationJobsResponse (Maybe Text)

The token to use to retreive the next page of results. This value is null when there are no more results to return.

listTextTranslationJobsResponse_httpStatus :: Lens' ListTextTranslationJobsResponse Int

The response's http status code.

StartTextTranslationJob

startTextTranslationJob_jobName :: Lens' StartTextTranslationJob (Maybe Text)

The name of the batch translation job to be performed.

startTextTranslationJob_parallelDataNames :: Lens' StartTextTranslationJob (Maybe [Text])

The names of the parallel data resources to use in the batch translation job. For a list of available parallel data resources, use the ListParallelData operation.

startTextTranslationJob_terminologyNames :: Lens' StartTextTranslationJob (Maybe [Text])

The name of the terminology to use in the batch translation job. For a list of available terminologies, use the ListTerminologies operation.

startTextTranslationJob_inputDataConfig :: Lens' StartTextTranslationJob InputDataConfig

Specifies the format and S3 location of the input documents for the translation job.

startTextTranslationJob_outputDataConfig :: Lens' StartTextTranslationJob OutputDataConfig

Specifies the S3 folder to which your job output will be saved.

startTextTranslationJob_dataAccessRoleArn :: Lens' StartTextTranslationJob Text

The Amazon Resource Name (ARN) of an AWS Identity Access and Management (IAM) role that grants Amazon Translate read access to your input data. For more nformation, see identity-and-access-management.

startTextTranslationJob_sourceLanguageCode :: Lens' StartTextTranslationJob Text

The language code of the input language. For a list of language codes, see what-is-languages.

Amazon Translate does not automatically detect a source language during batch translation jobs.

startTextTranslationJob_targetLanguageCodes :: Lens' StartTextTranslationJob (NonEmpty Text)

The language code of the output language.

startTextTranslationJob_clientToken :: Lens' StartTextTranslationJob Text

A unique identifier for the request. This token is auto-generated when using the Amazon Translate SDK.

startTextTranslationJobResponse_jobId :: Lens' StartTextTranslationJobResponse (Maybe Text)

The identifier generated for the job. To get the status of a job, use this ID with the DescribeTextTranslationJob operation.

startTextTranslationJobResponse_jobStatus :: Lens' StartTextTranslationJobResponse (Maybe JobStatus)

The status of the job. Possible values include:

SUBMITTED - The job has been received and is queued for processing.

IN_PROGRESS - Amazon Translate is processing the job.

COMPLETED - The job was successfully completed and the output is available.

COMPLETED_WITH_ERROR - The job was completed with errors. The errors can be analyzed in the job's output.

FAILED - The job did not complete. To get details, use the DescribeTextTranslationJob operation.

STOP_REQUESTED - The user who started the job has requested that it be stopped.

STOPPED - The job has been stopped.

startTextTranslationJobResponse_httpStatus :: Lens' StartTextTranslationJobResponse Int

The response's http status code.

ListParallelData

listParallelData_nextToken :: Lens' ListParallelData (Maybe Text)

A string that specifies the next page of results to return in a paginated response.

listParallelData_maxResults :: Lens' ListParallelData (Maybe Natural)

The maximum number of parallel data resources returned for each request.

listParallelDataResponse_parallelDataPropertiesList :: Lens' ListParallelDataResponse (Maybe [ParallelDataProperties])

The properties of the parallel data resources returned by this request.

listParallelDataResponse_nextToken :: Lens' ListParallelDataResponse (Maybe Text)

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

listParallelDataResponse_httpStatus :: Lens' ListParallelDataResponse Int

The response's http status code.

Types

AppliedTerminology

appliedTerminology_terms :: Lens' AppliedTerminology (Maybe [Term])

The specific terms of the custom terminology applied to the input text by Amazon Translate for the translated text response. A maximum of 250 terms will be returned, and the specific terms applied will be the first 250 terms in the source text.

appliedTerminology_name :: Lens' AppliedTerminology (Maybe Text)

The name of the custom terminology applied to the input text by Amazon Translate for the translated text response.

EncryptionKey

encryptionKey_type :: Lens' EncryptionKey EncryptionKeyType

The type of encryption key used by Amazon Translate to encrypt custom terminologies.

encryptionKey_id :: Lens' EncryptionKey Text

The Amazon Resource Name (ARN) of the encryption key being used to encrypt the custom terminology.

InputDataConfig

inputDataConfig_s3Uri :: Lens' InputDataConfig Text

The URI of the AWS S3 folder that contains the input file. The folder must be in the same Region as the API endpoint you are calling.

inputDataConfig_contentType :: Lens' InputDataConfig Text

Describes the format of the data that you submit to Amazon Translate as input. You can specify one of the following multipurpose internet mail extension (MIME) types:

text/html: The input data consists of one or more HTML files. Amazon Translate translates only the text that resides in the html element in each file.

text/plain: The input data consists of one or more unformatted text files. Amazon Translate translates every character in this type of input.

application/vnd.openxmlformats-officedocument.wordprocessingml.document: The input data consists of one or more Word documents (.docx).

application/vnd.openxmlformats-officedocument.presentationml.presentation: The input data consists of one or more PowerPoint Presentation files (.pptx).

application/vnd.openxmlformats-officedocument.spreadsheetml.sheet: The input data consists of one or more Excel Workbook files (.xlsx).

If you structure your input data as HTML, ensure that you set this parameter to text/html. By doing so, you cut costs by limiting the translation to the contents of the html element in each file. Otherwise, if you set this parameter to text/plain, your costs will cover the translation of every character.

JobDetails

jobDetails_translatedDocumentsCount :: Lens' JobDetails (Maybe Int)

The number of documents successfully processed during a translation job.

jobDetails_documentsWithErrorsCount :: Lens' JobDetails (Maybe Int)

The number of documents that could not be processed during a translation job.

jobDetails_inputDocumentsCount :: Lens' JobDetails (Maybe Int)

The number of documents used as input in a translation job.

OutputDataConfig

outputDataConfig_s3Uri :: Lens' OutputDataConfig Text

The URI of the S3 folder that contains a translation job's output file. The folder must be in the same Region as the API endpoint that you are calling.

ParallelDataConfig

parallelDataConfig_s3Uri :: Lens' ParallelDataConfig Text

The URI of the Amazon S3 folder that contains the parallel data input file. The folder must be in the same Region as the API endpoint you are calling.

parallelDataConfig_format :: Lens' ParallelDataConfig ParallelDataFormat

The format of the parallel data input file.

ParallelDataDataLocation

parallelDataDataLocation_repositoryType :: Lens' ParallelDataDataLocation Text

Describes the repository that contains the parallel data input file.

parallelDataDataLocation_location :: Lens' ParallelDataDataLocation Text

The Amazon S3 location of the parallel data input file. The location is returned as a presigned URL to that has a 30 minute expiration.

ParallelDataProperties

parallelDataProperties_status :: Lens' ParallelDataProperties (Maybe ParallelDataStatus)

The status of the parallel data resource. When the parallel data is ready for you to use, the status is ACTIVE.

parallelDataProperties_lastUpdatedAt :: Lens' ParallelDataProperties (Maybe UTCTime)

The time at which the parallel data resource was last updated.

parallelDataProperties_importedRecordCount :: Lens' ParallelDataProperties (Maybe Integer)

The number of records successfully imported from the parallel data input file.

parallelDataProperties_arn :: Lens' ParallelDataProperties (Maybe Text)

The Amazon Resource Name (ARN) of the parallel data resource.

parallelDataProperties_targetLanguageCodes :: Lens' ParallelDataProperties (Maybe [Text])

The language codes for the target languages available in the parallel data file. All possible target languages are returned as an array.

parallelDataProperties_createdAt :: Lens' ParallelDataProperties (Maybe UTCTime)

The time at which the parallel data resource was created.

parallelDataProperties_failedRecordCount :: Lens' ParallelDataProperties (Maybe Integer)

The number of records unsuccessfully imported from the parallel data input file.

parallelDataProperties_importedDataSize :: Lens' ParallelDataProperties (Maybe Integer)

The number of UTF-8 characters that Amazon Translate imported from the parallel data input file. This number includes only the characters in your translation examples. It does not include characters that are used to format your file. For example, if you provided a Translation Memory Exchange (.tmx) file, this number does not include the tags.

parallelDataProperties_name :: Lens' ParallelDataProperties (Maybe Text)

The custom name assigned to the parallel data resource.

parallelDataProperties_sourceLanguageCode :: Lens' ParallelDataProperties (Maybe Text)

The source language of the translations in the parallel data file.

parallelDataProperties_latestUpdateAttemptAt :: Lens' ParallelDataProperties (Maybe UTCTime)

The time that the most recent update was attempted.

parallelDataProperties_encryptionKey :: Lens' ParallelDataProperties (Maybe EncryptionKey)

Undocumented member.

parallelDataProperties_latestUpdateAttemptStatus :: Lens' ParallelDataProperties (Maybe ParallelDataStatus)

The status of the most recent update attempt for the parallel data resource.

parallelDataProperties_message :: Lens' ParallelDataProperties (Maybe Text)

Additional information from Amazon Translate about the parallel data resource.

parallelDataProperties_description :: Lens' ParallelDataProperties (Maybe Text)

The description assigned to the parallel data resource.

parallelDataProperties_skippedRecordCount :: Lens' ParallelDataProperties (Maybe Integer)

The number of items in the input file that Amazon Translate skipped when you created or updated the parallel data resource. For example, Amazon Translate skips empty records, empty target texts, and empty lines.

parallelDataProperties_parallelDataConfig :: Lens' ParallelDataProperties (Maybe ParallelDataConfig)

Specifies the format and S3 location of the parallel data input file.

Term

term_targetText :: Lens' Term (Maybe Text)

The target text of the term being translated by the custom terminology.

term_sourceText :: Lens' Term (Maybe Text)

The source text of the term being translated by the custom terminology.

TerminologyData

terminologyData_file :: Lens' TerminologyData ByteString

The file containing the custom terminology data. Your version of the AWS SDK performs a Base64-encoding on this field before sending a request to the AWS service. Users of the SDK should not perform Base64-encoding themselves.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

terminologyData_format :: Lens' TerminologyData TerminologyDataFormat

The data format of the custom terminology. Either CSV or TMX.

TerminologyDataLocation

terminologyDataLocation_repositoryType :: Lens' TerminologyDataLocation Text

The repository type for the custom terminology data.

terminologyDataLocation_location :: Lens' TerminologyDataLocation Text

The location of the custom terminology data.

TerminologyProperties

terminologyProperties_sizeBytes :: Lens' TerminologyProperties (Maybe Int)

The size of the file used when importing a custom terminology.

terminologyProperties_lastUpdatedAt :: Lens' TerminologyProperties (Maybe UTCTime)

The time at which the custom terminology was last update, based on the timestamp.

terminologyProperties_arn :: Lens' TerminologyProperties (Maybe Text)

The Amazon Resource Name (ARN) of the custom terminology.

terminologyProperties_targetLanguageCodes :: Lens' TerminologyProperties (Maybe [Text])

The language codes for the target languages available with the custom terminology file. All possible target languages are returned in array.

terminologyProperties_createdAt :: Lens' TerminologyProperties (Maybe UTCTime)

The time at which the custom terminology was created, based on the timestamp.

terminologyProperties_name :: Lens' TerminologyProperties (Maybe Text)

The name of the custom terminology.

terminologyProperties_sourceLanguageCode :: Lens' TerminologyProperties (Maybe Text)

The language code for the source text of the translation request for which the custom terminology is being used.

terminologyProperties_termCount :: Lens' TerminologyProperties (Maybe Int)

The number of terms included in the custom terminology.

terminologyProperties_encryptionKey :: Lens' TerminologyProperties (Maybe EncryptionKey)

The encryption key for the custom terminology.

terminologyProperties_description :: Lens' TerminologyProperties (Maybe Text)

The description of the custom terminology properties.

TextTranslationJobFilter

textTranslationJobFilter_submittedBeforeTime :: Lens' TextTranslationJobFilter (Maybe UTCTime)

Filters the list of jobs based on the time that the job was submitted for processing and returns only the jobs submitted before the specified time. Jobs are returned in ascending order, oldest to newest.

textTranslationJobFilter_submittedAfterTime :: Lens' TextTranslationJobFilter (Maybe UTCTime)

Filters the list of jobs based on the time that the job was submitted for processing and returns only the jobs submitted after the specified time. Jobs are returned in descending order, newest to oldest.

textTranslationJobFilter_jobName :: Lens' TextTranslationJobFilter (Maybe Text)

Filters the list of jobs by name.

textTranslationJobFilter_jobStatus :: Lens' TextTranslationJobFilter (Maybe JobStatus)

Filters the list of jobs based by job status.

TextTranslationJobProperties

textTranslationJobProperties_jobId :: Lens' TextTranslationJobProperties (Maybe Text)

The ID of the translation job.

textTranslationJobProperties_targetLanguageCodes :: Lens' TextTranslationJobProperties (Maybe (NonEmpty Text))

The language code of the language of the target text. The language must be a language supported by Amazon Translate.

textTranslationJobProperties_jobName :: Lens' TextTranslationJobProperties (Maybe Text)

The user-defined name of the translation job.

textTranslationJobProperties_submittedTime :: Lens' TextTranslationJobProperties (Maybe UTCTime)

The time at which the translation job was submitted.

textTranslationJobProperties_inputDataConfig :: Lens' TextTranslationJobProperties (Maybe InputDataConfig)

The input configuration properties that were specified when the job was requested.

textTranslationJobProperties_parallelDataNames :: Lens' TextTranslationJobProperties (Maybe [Text])

A list containing the names of the parallel data resources applied to the translation job.

textTranslationJobProperties_terminologyNames :: Lens' TextTranslationJobProperties (Maybe [Text])

A list containing the names of the terminologies applied to a translation job. Only one terminology can be applied per StartTextTranslationJob request at this time.

textTranslationJobProperties_sourceLanguageCode :: Lens' TextTranslationJobProperties (Maybe Text)

The language code of the language of the source text. The language must be a language supported by Amazon Translate.

textTranslationJobProperties_endTime :: Lens' TextTranslationJobProperties (Maybe UTCTime)

The time at which the translation job ended.

textTranslationJobProperties_outputDataConfig :: Lens' TextTranslationJobProperties (Maybe OutputDataConfig)

The output configuration properties that were specified when the job was requested.

textTranslationJobProperties_jobDetails :: Lens' TextTranslationJobProperties (Maybe JobDetails)

The number of documents successfully and unsuccessfully processed during the translation job.

textTranslationJobProperties_dataAccessRoleArn :: Lens' TextTranslationJobProperties (Maybe Text)

The Amazon Resource Name (ARN) of an AWS Identity Access and Management (IAM) role that granted Amazon Translate read access to the job's input data.

textTranslationJobProperties_jobStatus :: Lens' TextTranslationJobProperties (Maybe JobStatus)

The status of the translation job.

textTranslationJobProperties_message :: Lens' TextTranslationJobProperties (Maybe Text)

An explanation of any errors that may have occured during the translation job.