Network.AWS.Translate.ListTextTranslationJobs

Description

Gets a list of the batch translation jobs that you have submitted.

Creating a Request

data ListTextTranslationJobs

See: newListTextTranslationJobs smart constructor.

Constructors

ListTextTranslationJobs'

Fields

nextToken :: Maybe Text

The token to request the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return in each page. The default value is 100.

filter' :: Maybe TextTranslationJobFilter

The parameters that specify which batch translation jobs to retrieve. Filters include job name, job status, and submission time. You can only set one filter at a time.

newListTextTranslationJobs :: ListTextTranslationJobs

Create a value of ListTextTranslationJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTextTranslationJobs', listTextTranslationJobs_nextToken - The token to request the next page of results.

$sel:maxResults:ListTextTranslationJobs', listTextTranslationJobs_maxResults - The maximum number of results to return in each page. The default value is 100.

$sel:filter':ListTextTranslationJobs', listTextTranslationJobs_filter - The parameters that specify which batch translation jobs to retrieve. Filters include job name, job status, and submission time. You can only set one filter at a time.

Request Lenses

listTextTranslationJobs_nextToken :: Lens' ListTextTranslationJobs (Maybe Text)

The token to request the next page of results.

listTextTranslationJobs_maxResults :: Lens' ListTextTranslationJobs (Maybe Natural)

The maximum number of results to return in each page. The default value is 100.

listTextTranslationJobs_filter :: Lens' ListTextTranslationJobs (Maybe TextTranslationJobFilter)

The parameters that specify which batch translation jobs to retrieve. Filters include job name, job status, and submission time. You can only set one filter at a time.

Destructuring the Response

data ListTextTranslationJobsResponse

See: newListTextTranslationJobsResponse smart constructor.

Constructors

ListTextTranslationJobsResponse'

Fields

nextToken :: Maybe Text

The token to use to retreive the next page of results. This value is null when there are no more results to return.

textTranslationJobPropertiesList :: Maybe [TextTranslationJobProperties]

A list containing the properties of each job that is returned.

httpStatus :: Int

The response's http status code.

newListTextTranslationJobsResponse

Arguments

:: Int
$sel:httpStatus:ListTextTranslationJobsResponse'

-> ListTextTranslationJobsResponse

Create a value of ListTextTranslationJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTextTranslationJobs', listTextTranslationJobsResponse_nextToken - The token to use to retreive the next page of results. This value is null when there are no more results to return.

$sel:textTranslationJobPropertiesList:ListTextTranslationJobsResponse', listTextTranslationJobsResponse_textTranslationJobPropertiesList - A list containing the properties of each job that is returned.

$sel:httpStatus:ListTextTranslationJobsResponse', listTextTranslationJobsResponse_httpStatus - The response's http status code.

Response Lenses

listTextTranslationJobsResponse_nextToken :: Lens' ListTextTranslationJobsResponse (Maybe Text)

The token to use to retreive the next page of results. This value is null when there are no more results to return.

listTextTranslationJobsResponse_textTranslationJobPropertiesList :: Lens' ListTextTranslationJobsResponse (Maybe [TextTranslationJobProperties])

A list containing the properties of each job that is returned.

listTextTranslationJobsResponse_httpStatus :: Lens' ListTextTranslationJobsResponse Int

The response's http status code.