Network.AWS.Translate.TranslateText

Description

Translates input text from the source language to the target language. For a list of available languages and language codes, see what-is-languages.

Creating a Request

data TranslateText

See: newTranslateText smart constructor.

Constructors

TranslateText'

Fields

terminologyNames :: Maybe [Text]

The name of the terminology list file to be used in the TranslateText request. You can use 1 terminology list at most in a TranslateText request. Terminology lists can contain a maximum of 256 terms.

text :: Text

The text to translate. The text string can be a maximum of 5,000 bytes long. Depending on your character set, this may be fewer than 5,000 characters.

sourceLanguageCode :: Text

The language code for the language of the source text. The language must be a language supported by Amazon Translate. For a list of language codes, see what-is-languages.

To have Amazon Translate determine the source language of your text, you can specify auto in the SourceLanguageCode field. If you specify auto, Amazon Translate will call Amazon Comprehend to determine the source language.

targetLanguageCode :: Text

The language code requested for the language of the target text. The language must be a language supported by Amazon Translate.

newTranslateText

Arguments

:: Text
$sel:text:TranslateText'

-> Text
$sel:sourceLanguageCode:TranslateText'

-> Text
$sel:targetLanguageCode:TranslateText'

-> TranslateText

Create a value of TranslateText with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:terminologyNames:TranslateText', translateText_terminologyNames - The name of the terminology list file to be used in the TranslateText request. You can use 1 terminology list at most in a TranslateText request. Terminology lists can contain a maximum of 256 terms.

$sel:text:TranslateText', translateText_text - The text to translate. The text string can be a maximum of 5,000 bytes long. Depending on your character set, this may be fewer than 5,000 characters.

$sel:sourceLanguageCode:TranslateText', translateText_sourceLanguageCode - The language code for the language of the source text. The language must be a language supported by Amazon Translate. For a list of language codes, see what-is-languages.

To have Amazon Translate determine the source language of your text, you can specify auto in the SourceLanguageCode field. If you specify auto, Amazon Translate will call Amazon Comprehend to determine the source language.

$sel:targetLanguageCode:TranslateText', translateText_targetLanguageCode - The language code requested for the language of the target text. The language must be a language supported by Amazon Translate.

Request Lenses

translateText_terminologyNames :: Lens' TranslateText (Maybe [Text])

The name of the terminology list file to be used in the TranslateText request. You can use 1 terminology list at most in a TranslateText request. Terminology lists can contain a maximum of 256 terms.

translateText_text :: Lens' TranslateText Text

The text to translate. The text string can be a maximum of 5,000 bytes long. Depending on your character set, this may be fewer than 5,000 characters.

translateText_sourceLanguageCode :: Lens' TranslateText Text

The language code for the language of the source text. The language must be a language supported by Amazon Translate. For a list of language codes, see what-is-languages.

To have Amazon Translate determine the source language of your text, you can specify auto in the SourceLanguageCode field. If you specify auto, Amazon Translate will call Amazon Comprehend to determine the source language.

translateText_targetLanguageCode :: Lens' TranslateText Text

The language code requested for the language of the target text. The language must be a language supported by Amazon Translate.

Destructuring the Response

data TranslateTextResponse

See: newTranslateTextResponse smart constructor.

Constructors

TranslateTextResponse'

Fields

appliedTerminologies :: Maybe [AppliedTerminology]

The names of the custom terminologies applied to the input text by Amazon Translate for the translated text response.

httpStatus :: Int

The response's http status code.

translatedText :: Text

The translated text.

sourceLanguageCode :: Text

The language code for the language of the source text.

targetLanguageCode :: Text

The language code for the language of the target text.

newTranslateTextResponse

Arguments

:: Int
$sel:httpStatus:TranslateTextResponse'

-> Text
$sel:translatedText:TranslateTextResponse'

-> Text
$sel:sourceLanguageCode:TranslateText'

-> Text
$sel:targetLanguageCode:TranslateText'

-> TranslateTextResponse

Create a value of TranslateTextResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appliedTerminologies:TranslateTextResponse', translateTextResponse_appliedTerminologies - The names of the custom terminologies applied to the input text by Amazon Translate for the translated text response.

$sel:httpStatus:TranslateTextResponse', translateTextResponse_httpStatus - The response's http status code.

$sel:translatedText:TranslateTextResponse', translateTextResponse_translatedText - The translated text.

$sel:sourceLanguageCode:TranslateText', translateTextResponse_sourceLanguageCode - The language code for the language of the source text.

$sel:targetLanguageCode:TranslateText', translateTextResponse_targetLanguageCode - The language code for the language of the target text.

Response Lenses

translateTextResponse_appliedTerminologies :: Lens' TranslateTextResponse (Maybe [AppliedTerminology])

The names of the custom terminologies applied to the input text by Amazon Translate for the translated text response.

translateTextResponse_httpStatus :: Lens' TranslateTextResponse Int

The response's http status code.

translateTextResponse_translatedText :: Lens' TranslateTextResponse Text

The translated text.

translateTextResponse_sourceLanguageCode :: Lens' TranslateTextResponse Text

The language code for the language of the source text.

translateTextResponse_targetLanguageCode :: Lens' TranslateTextResponse Text

The language code for the language of the target text.