Network.AWS.Translate.Types

Description

Service Configuration

defaultService :: Service

API version 2017-07-01 of the Amazon Translate SDK configuration.

Errors

_InvalidFilterException :: AsError a => Getting (First ServiceError) a ServiceError

The filter specified for the operation is invalid. Specify a different filter.

_DetectedLanguageLowConfidenceException :: AsError a => Getting (First ServiceError) a ServiceError

The confidence that Amazon Comprehend accurately detected the source language is low. If a low confidence level is acceptable for your application, you can use the language in the exception to call Amazon Translate again. For more information, see the DetectDominantLanguage operation in the Amazon Comprehend Developer Guide.

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The Amazon Translate service is temporarily unavailable. Please wait a bit and then retry your request.

_UnsupportedLanguagePairException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Translate does not support translation from the language of the source text into the requested target language. For more information, see how-to-error-msg.

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

Another modification is being made. That modification must complete before you can make your change.

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request that you made is invalid. Check your request to determine why it's invalid and then retry the request.

_InvalidParameterValueException :: AsError a => Getting (First ServiceError) a ServiceError

The value of the parameter is invalid. Review the value of the parameter you are using to correct it, and then retry your operation.

_TextSizeLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The size of the text you submitted exceeds the size limit. Reduce the size of the text or use a smaller document and then retry your request.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

There was a conflict processing the request. Try your request again.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The specified limit has been exceeded. Review your request and retry it with a quantity below the stated limit.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resource you are looking for has not been found. Review the resource you're looking for and see if a different resource will accomplish your needs before retrying the revised request.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

An internal server error occurred. Retry your request.

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

You have made too many requests within a short period of time. Wait for a short time and then try your request again.

EncryptionKeyType

newtype EncryptionKeyType

Constructors

EncryptionKeyType'

Fields

fromEncryptionKeyType :: Text

Bundled Patterns

pattern EncryptionKeyType_KMS :: EncryptionKeyType

JobStatus

newtype JobStatus

Constructors

JobStatus'

Fields

fromJobStatus :: Text

Bundled Patterns

pattern JobStatus_COMPLETED :: JobStatus

pattern JobStatus_COMPLETED_WITH_ERROR :: JobStatus

pattern JobStatus_FAILED :: JobStatus

pattern JobStatus_IN_PROGRESS :: JobStatus

pattern JobStatus_STOPPED :: JobStatus

pattern JobStatus_STOP_REQUESTED :: JobStatus

pattern JobStatus_SUBMITTED :: JobStatus

MergeStrategy

newtype MergeStrategy

Constructors

MergeStrategy'

Fields

fromMergeStrategy :: Text

Bundled Patterns

pattern MergeStrategy_OVERWRITE :: MergeStrategy

ParallelDataFormat

newtype ParallelDataFormat

Constructors

ParallelDataFormat'

Fields

fromParallelDataFormat :: Text

Bundled Patterns

pattern ParallelDataFormat_CSV :: ParallelDataFormat

pattern ParallelDataFormat_TMX :: ParallelDataFormat

pattern ParallelDataFormat_TSV :: ParallelDataFormat

ParallelDataStatus

newtype ParallelDataStatus

Constructors

ParallelDataStatus'

Fields

fromParallelDataStatus :: Text

Bundled Patterns

pattern ParallelDataStatus_ACTIVE :: ParallelDataStatus

pattern ParallelDataStatus_CREATING :: ParallelDataStatus

pattern ParallelDataStatus_DELETING :: ParallelDataStatus

pattern ParallelDataStatus_FAILED :: ParallelDataStatus

pattern ParallelDataStatus_UPDATING :: ParallelDataStatus

TerminologyDataFormat

newtype TerminologyDataFormat

Constructors

TerminologyDataFormat'

Fields

fromTerminologyDataFormat :: Text

Bundled Patterns

pattern TerminologyDataFormat_CSV :: TerminologyDataFormat

pattern TerminologyDataFormat_TMX :: TerminologyDataFormat

AppliedTerminology

data AppliedTerminology

The custom terminology applied to the input text by Amazon Translate for the translated text response. This is optional in the response and will only be present if you specified terminology input in the request. Currently, only one terminology can be applied per TranslateText request.

See: newAppliedTerminology smart constructor.

Constructors

AppliedTerminology'

Fields

terms :: Maybe [Term]

The specific terms of the custom terminology applied to the input text by Amazon Translate for the translated text response. A maximum of 250 terms will be returned, and the specific terms applied will be the first 250 terms in the source text.

name :: Maybe Text

The name of the custom terminology applied to the input text by Amazon Translate for the translated text response.

newAppliedTerminology :: AppliedTerminology

Create a value of AppliedTerminology with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:terms:AppliedTerminology', appliedTerminology_terms - The specific terms of the custom terminology applied to the input text by Amazon Translate for the translated text response. A maximum of 250 terms will be returned, and the specific terms applied will be the first 250 terms in the source text.

$sel:name:AppliedTerminology', appliedTerminology_name - The name of the custom terminology applied to the input text by Amazon Translate for the translated text response.

appliedTerminology_terms :: Lens' AppliedTerminology (Maybe [Term])

The specific terms of the custom terminology applied to the input text by Amazon Translate for the translated text response. A maximum of 250 terms will be returned, and the specific terms applied will be the first 250 terms in the source text.

appliedTerminology_name :: Lens' AppliedTerminology (Maybe Text)

The name of the custom terminology applied to the input text by Amazon Translate for the translated text response.

EncryptionKey

data EncryptionKey

The encryption key used to encrypt this object.

See: newEncryptionKey smart constructor.

Constructors

EncryptionKey'

Fields

type' :: EncryptionKeyType

The type of encryption key used by Amazon Translate to encrypt custom terminologies.

id :: Text

The Amazon Resource Name (ARN) of the encryption key being used to encrypt the custom terminology.

newEncryptionKey

Arguments

:: EncryptionKeyType
$sel:type':EncryptionKey'

-> Text
$sel:id:EncryptionKey'

-> EncryptionKey

Create a value of EncryptionKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':EncryptionKey', encryptionKey_type - The type of encryption key used by Amazon Translate to encrypt custom terminologies.

$sel:id:EncryptionKey', encryptionKey_id - The Amazon Resource Name (ARN) of the encryption key being used to encrypt the custom terminology.

encryptionKey_type :: Lens' EncryptionKey EncryptionKeyType

The type of encryption key used by Amazon Translate to encrypt custom terminologies.

encryptionKey_id :: Lens' EncryptionKey Text

The Amazon Resource Name (ARN) of the encryption key being used to encrypt the custom terminology.

InputDataConfig

data InputDataConfig

The input configuration properties for requesting a batch translation job.

See: newInputDataConfig smart constructor.

Constructors

InputDataConfig'

Fields

s3Uri :: Text

The URI of the AWS S3 folder that contains the input file. The folder must be in the same Region as the API endpoint you are calling.

contentType :: Text

Describes the format of the data that you submit to Amazon Translate as input. You can specify one of the following multipurpose internet mail extension (MIME) types:

text/html: The input data consists of one or more HTML files. Amazon Translate translates only the text that resides in the html element in each file.

text/plain: The input data consists of one or more unformatted text files. Amazon Translate translates every character in this type of input.

application/vnd.openxmlformats-officedocument.wordprocessingml.document: The input data consists of one or more Word documents (.docx).

application/vnd.openxmlformats-officedocument.presentationml.presentation: The input data consists of one or more PowerPoint Presentation files (.pptx).

application/vnd.openxmlformats-officedocument.spreadsheetml.sheet: The input data consists of one or more Excel Workbook files (.xlsx).

If you structure your input data as HTML, ensure that you set this parameter to text/html. By doing so, you cut costs by limiting the translation to the contents of the html element in each file. Otherwise, if you set this parameter to text/plain, your costs will cover the translation of every character.

newInputDataConfig

Arguments

:: Text
$sel:s3Uri:InputDataConfig'

-> Text
$sel:contentType:InputDataConfig'

-> InputDataConfig

Create a value of InputDataConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Uri:InputDataConfig', inputDataConfig_s3Uri - The URI of the AWS S3 folder that contains the input file. The folder must be in the same Region as the API endpoint you are calling.

$sel:contentType:InputDataConfig', inputDataConfig_contentType - Describes the format of the data that you submit to Amazon Translate as input. You can specify one of the following multipurpose internet mail extension (MIME) types:

text/html: The input data consists of one or more HTML files. Amazon Translate translates only the text that resides in the html element in each file.

text/plain: The input data consists of one or more unformatted text files. Amazon Translate translates every character in this type of input.

application/vnd.openxmlformats-officedocument.wordprocessingml.document: The input data consists of one or more Word documents (.docx).

application/vnd.openxmlformats-officedocument.presentationml.presentation: The input data consists of one or more PowerPoint Presentation files (.pptx).

application/vnd.openxmlformats-officedocument.spreadsheetml.sheet: The input data consists of one or more Excel Workbook files (.xlsx).

If you structure your input data as HTML, ensure that you set this parameter to text/html. By doing so, you cut costs by limiting the translation to the contents of the html element in each file. Otherwise, if you set this parameter to text/plain, your costs will cover the translation of every character.

inputDataConfig_s3Uri :: Lens' InputDataConfig Text

The URI of the AWS S3 folder that contains the input file. The folder must be in the same Region as the API endpoint you are calling.

inputDataConfig_contentType :: Lens' InputDataConfig Text

Describes the format of the data that you submit to Amazon Translate as input. You can specify one of the following multipurpose internet mail extension (MIME) types:

text/html: The input data consists of one or more HTML files. Amazon Translate translates only the text that resides in the html element in each file.

text/plain: The input data consists of one or more unformatted text files. Amazon Translate translates every character in this type of input.

application/vnd.openxmlformats-officedocument.wordprocessingml.document: The input data consists of one or more Word documents (.docx).

application/vnd.openxmlformats-officedocument.presentationml.presentation: The input data consists of one or more PowerPoint Presentation files (.pptx).

application/vnd.openxmlformats-officedocument.spreadsheetml.sheet: The input data consists of one or more Excel Workbook files (.xlsx).

If you structure your input data as HTML, ensure that you set this parameter to text/html. By doing so, you cut costs by limiting the translation to the contents of the html element in each file. Otherwise, if you set this parameter to text/plain, your costs will cover the translation of every character.

JobDetails

data JobDetails

The number of documents successfully and unsuccessfully processed during a translation job.

See: newJobDetails smart constructor.

Constructors

JobDetails'

Fields

inputDocumentsCount :: Maybe Int

The number of documents used as input in a translation job.

documentsWithErrorsCount :: Maybe Int

The number of documents that could not be processed during a translation job.

translatedDocumentsCount :: Maybe Int

The number of documents successfully processed during a translation job.

newJobDetails :: JobDetails

Create a value of JobDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputDocumentsCount:JobDetails', jobDetails_inputDocumentsCount - The number of documents used as input in a translation job.

$sel:documentsWithErrorsCount:JobDetails', jobDetails_documentsWithErrorsCount - The number of documents that could not be processed during a translation job.

$sel:translatedDocumentsCount:JobDetails', jobDetails_translatedDocumentsCount - The number of documents successfully processed during a translation job.

jobDetails_inputDocumentsCount :: Lens' JobDetails (Maybe Int)

The number of documents used as input in a translation job.

jobDetails_documentsWithErrorsCount :: Lens' JobDetails (Maybe Int)

The number of documents that could not be processed during a translation job.

jobDetails_translatedDocumentsCount :: Lens' JobDetails (Maybe Int)

The number of documents successfully processed during a translation job.

OutputDataConfig

data OutputDataConfig

The output configuration properties for a batch translation job.

See: newOutputDataConfig smart constructor.

Constructors

OutputDataConfig'

Fields

s3Uri :: Text

The URI of the S3 folder that contains a translation job's output file. The folder must be in the same Region as the API endpoint that you are calling.

newOutputDataConfig

Arguments

:: Text
$sel:s3Uri:OutputDataConfig'

-> OutputDataConfig

Create a value of OutputDataConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Uri:OutputDataConfig', outputDataConfig_s3Uri - The URI of the S3 folder that contains a translation job's output file. The folder must be in the same Region as the API endpoint that you are calling.

outputDataConfig_s3Uri :: Lens' OutputDataConfig Text

The URI of the S3 folder that contains a translation job's output file. The folder must be in the same Region as the API endpoint that you are calling.

ParallelDataConfig

data ParallelDataConfig

Specifies the format and S3 location of the parallel data input file.

See: newParallelDataConfig smart constructor.

Constructors

ParallelDataConfig'

Fields

s3Uri :: Text

The URI of the Amazon S3 folder that contains the parallel data input file. The folder must be in the same Region as the API endpoint you are calling.

format :: ParallelDataFormat

The format of the parallel data input file.

newParallelDataConfig

Arguments

:: Text
$sel:s3Uri:ParallelDataConfig'

-> ParallelDataFormat
$sel:format:ParallelDataConfig'

-> ParallelDataConfig

Create a value of ParallelDataConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Uri:ParallelDataConfig', parallelDataConfig_s3Uri - The URI of the Amazon S3 folder that contains the parallel data input file. The folder must be in the same Region as the API endpoint you are calling.

$sel:format:ParallelDataConfig', parallelDataConfig_format - The format of the parallel data input file.

parallelDataConfig_s3Uri :: Lens' ParallelDataConfig Text

The URI of the Amazon S3 folder that contains the parallel data input file. The folder must be in the same Region as the API endpoint you are calling.

parallelDataConfig_format :: Lens' ParallelDataConfig ParallelDataFormat

The format of the parallel data input file.

ParallelDataDataLocation

data ParallelDataDataLocation

The location of the most recent parallel data input file that was successfully imported into Amazon Translate.

See: newParallelDataDataLocation smart constructor.

Constructors

ParallelDataDataLocation'

Fields

repositoryType :: Text

Describes the repository that contains the parallel data input file.

location :: Text

The Amazon S3 location of the parallel data input file. The location is returned as a presigned URL to that has a 30 minute expiration.

newParallelDataDataLocation

Arguments

:: Text
$sel:repositoryType:ParallelDataDataLocation'

-> Text
$sel:location:ParallelDataDataLocation'

-> ParallelDataDataLocation

Create a value of ParallelDataDataLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositoryType:ParallelDataDataLocation', parallelDataDataLocation_repositoryType - Describes the repository that contains the parallel data input file.

$sel:location:ParallelDataDataLocation', parallelDataDataLocation_location - The Amazon S3 location of the parallel data input file. The location is returned as a presigned URL to that has a 30 minute expiration.

parallelDataDataLocation_repositoryType :: Lens' ParallelDataDataLocation Text

Describes the repository that contains the parallel data input file.

parallelDataDataLocation_location :: Lens' ParallelDataDataLocation Text

The Amazon S3 location of the parallel data input file. The location is returned as a presigned URL to that has a 30 minute expiration.

ParallelDataProperties

data ParallelDataProperties

The properties of a parallel data resource.

See: newParallelDataProperties smart constructor.

Constructors

ParallelDataProperties'

Fields

status :: Maybe ParallelDataStatus

The status of the parallel data resource. When the parallel data is ready for you to use, the status is ACTIVE.

importedDataSize :: Maybe Integer

The number of UTF-8 characters that Amazon Translate imported from the parallel data input file. This number includes only the characters in your translation examples. It does not include characters that are used to format your file. For example, if you provided a Translation Memory Exchange (.tmx) file, this number does not include the tags.

skippedRecordCount :: Maybe Integer

The number of items in the input file that Amazon Translate skipped when you created or updated the parallel data resource. For example, Amazon Translate skips empty records, empty target texts, and empty lines.

latestUpdateAttemptStatus :: Maybe ParallelDataStatus

The status of the most recent update attempt for the parallel data resource.

message :: Maybe Text

Additional information from Amazon Translate about the parallel data resource.

encryptionKey :: Maybe EncryptionKey

createdAt :: Maybe POSIX

The time at which the parallel data resource was created.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the parallel data resource.

failedRecordCount :: Maybe Integer

The number of records unsuccessfully imported from the parallel data input file.

targetLanguageCodes :: Maybe [Text]

The language codes for the target languages available in the parallel data file. All possible target languages are returned as an array.

latestUpdateAttemptAt :: Maybe POSIX

The time that the most recent update was attempted.

name :: Maybe Text

The custom name assigned to the parallel data resource.

parallelDataConfig :: Maybe ParallelDataConfig

Specifies the format and S3 location of the parallel data input file.

description :: Maybe Text

The description assigned to the parallel data resource.

sourceLanguageCode :: Maybe Text

The source language of the translations in the parallel data file.

lastUpdatedAt :: Maybe POSIX

The time at which the parallel data resource was last updated.

importedRecordCount :: Maybe Integer

The number of records successfully imported from the parallel data input file.

newParallelDataProperties :: ParallelDataProperties

Create a value of ParallelDataProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ParallelDataProperties', parallelDataProperties_status - The status of the parallel data resource. When the parallel data is ready for you to use, the status is ACTIVE.

$sel:importedDataSize:ParallelDataProperties', parallelDataProperties_importedDataSize - The number of UTF-8 characters that Amazon Translate imported from the parallel data input file. This number includes only the characters in your translation examples. It does not include characters that are used to format your file. For example, if you provided a Translation Memory Exchange (.tmx) file, this number does not include the tags.

$sel:skippedRecordCount:ParallelDataProperties', parallelDataProperties_skippedRecordCount - The number of items in the input file that Amazon Translate skipped when you created or updated the parallel data resource. For example, Amazon Translate skips empty records, empty target texts, and empty lines.

$sel:latestUpdateAttemptStatus:ParallelDataProperties', parallelDataProperties_latestUpdateAttemptStatus - The status of the most recent update attempt for the parallel data resource.

$sel:message:ParallelDataProperties', parallelDataProperties_message - Additional information from Amazon Translate about the parallel data resource.

$sel:encryptionKey:ParallelDataProperties', parallelDataProperties_encryptionKey - Undocumented member.

$sel:createdAt:ParallelDataProperties', parallelDataProperties_createdAt - The time at which the parallel data resource was created.

$sel:arn:ParallelDataProperties', parallelDataProperties_arn - The Amazon Resource Name (ARN) of the parallel data resource.

$sel:failedRecordCount:ParallelDataProperties', parallelDataProperties_failedRecordCount - The number of records unsuccessfully imported from the parallel data input file.

$sel:targetLanguageCodes:ParallelDataProperties', parallelDataProperties_targetLanguageCodes - The language codes for the target languages available in the parallel data file. All possible target languages are returned as an array.

$sel:latestUpdateAttemptAt:ParallelDataProperties', parallelDataProperties_latestUpdateAttemptAt - The time that the most recent update was attempted.

$sel:name:ParallelDataProperties', parallelDataProperties_name - The custom name assigned to the parallel data resource.

$sel:parallelDataConfig:ParallelDataProperties', parallelDataProperties_parallelDataConfig - Specifies the format and S3 location of the parallel data input file.

$sel:description:ParallelDataProperties', parallelDataProperties_description - The description assigned to the parallel data resource.

$sel:sourceLanguageCode:ParallelDataProperties', parallelDataProperties_sourceLanguageCode - The source language of the translations in the parallel data file.

$sel:lastUpdatedAt:ParallelDataProperties', parallelDataProperties_lastUpdatedAt - The time at which the parallel data resource was last updated.

$sel:importedRecordCount:ParallelDataProperties', parallelDataProperties_importedRecordCount - The number of records successfully imported from the parallel data input file.

parallelDataProperties_status :: Lens' ParallelDataProperties (Maybe ParallelDataStatus)

The status of the parallel data resource. When the parallel data is ready for you to use, the status is ACTIVE.

parallelDataProperties_importedDataSize :: Lens' ParallelDataProperties (Maybe Integer)

The number of UTF-8 characters that Amazon Translate imported from the parallel data input file. This number includes only the characters in your translation examples. It does not include characters that are used to format your file. For example, if you provided a Translation Memory Exchange (.tmx) file, this number does not include the tags.

parallelDataProperties_skippedRecordCount :: Lens' ParallelDataProperties (Maybe Integer)

The number of items in the input file that Amazon Translate skipped when you created or updated the parallel data resource. For example, Amazon Translate skips empty records, empty target texts, and empty lines.

parallelDataProperties_latestUpdateAttemptStatus :: Lens' ParallelDataProperties (Maybe ParallelDataStatus)

The status of the most recent update attempt for the parallel data resource.

parallelDataProperties_message :: Lens' ParallelDataProperties (Maybe Text)

Additional information from Amazon Translate about the parallel data resource.

parallelDataProperties_encryptionKey :: Lens' ParallelDataProperties (Maybe EncryptionKey)

Undocumented member.

parallelDataProperties_createdAt :: Lens' ParallelDataProperties (Maybe UTCTime)

The time at which the parallel data resource was created.

parallelDataProperties_arn :: Lens' ParallelDataProperties (Maybe Text)

The Amazon Resource Name (ARN) of the parallel data resource.

parallelDataProperties_failedRecordCount :: Lens' ParallelDataProperties (Maybe Integer)

The number of records unsuccessfully imported from the parallel data input file.

parallelDataProperties_targetLanguageCodes :: Lens' ParallelDataProperties (Maybe [Text])

The language codes for the target languages available in the parallel data file. All possible target languages are returned as an array.

parallelDataProperties_latestUpdateAttemptAt :: Lens' ParallelDataProperties (Maybe UTCTime)

The time that the most recent update was attempted.

parallelDataProperties_name :: Lens' ParallelDataProperties (Maybe Text)

The custom name assigned to the parallel data resource.

parallelDataProperties_parallelDataConfig :: Lens' ParallelDataProperties (Maybe ParallelDataConfig)

Specifies the format and S3 location of the parallel data input file.

parallelDataProperties_description :: Lens' ParallelDataProperties (Maybe Text)

The description assigned to the parallel data resource.

parallelDataProperties_sourceLanguageCode :: Lens' ParallelDataProperties (Maybe Text)

The source language of the translations in the parallel data file.

parallelDataProperties_lastUpdatedAt :: Lens' ParallelDataProperties (Maybe UTCTime)

The time at which the parallel data resource was last updated.

parallelDataProperties_importedRecordCount :: Lens' ParallelDataProperties (Maybe Integer)

The number of records successfully imported from the parallel data input file.

Term

data Term

The term being translated by the custom terminology.

See: newTerm smart constructor.

Constructors

Term'

Fields

targetText :: Maybe Text

The target text of the term being translated by the custom terminology.

sourceText :: Maybe Text

The source text of the term being translated by the custom terminology.

newTerm :: Term

Create a value of Term with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetText:Term', term_targetText - The target text of the term being translated by the custom terminology.

$sel:sourceText:Term', term_sourceText - The source text of the term being translated by the custom terminology.

term_targetText :: Lens' Term (Maybe Text)

The target text of the term being translated by the custom terminology.

term_sourceText :: Lens' Term (Maybe Text)

The source text of the term being translated by the custom terminology.

TerminologyData

data TerminologyData

The data associated with the custom terminology.

See: newTerminologyData smart constructor.

Constructors

TerminologyData'

Fields

file :: Sensitive Base64

The file containing the custom terminology data. Your version of the AWS SDK performs a Base64-encoding on this field before sending a request to the AWS service. Users of the SDK should not perform Base64-encoding themselves.

format :: TerminologyDataFormat

The data format of the custom terminology. Either CSV or TMX.

newTerminologyData

Arguments

:: ByteString
$sel:file:TerminologyData'

-> TerminologyDataFormat
$sel:format:TerminologyData'

-> TerminologyData

Create a value of TerminologyData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:file:TerminologyData', terminologyData_file - The file containing the custom terminology data. Your version of the AWS SDK performs a Base64-encoding on this field before sending a request to the AWS service. Users of the SDK should not perform Base64-encoding themselves.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:format:TerminologyData', terminologyData_format - The data format of the custom terminology. Either CSV or TMX.

terminologyData_file :: Lens' TerminologyData ByteString

The file containing the custom terminology data. Your version of the AWS SDK performs a Base64-encoding on this field before sending a request to the AWS service. Users of the SDK should not perform Base64-encoding themselves.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

terminologyData_format :: Lens' TerminologyData TerminologyDataFormat

The data format of the custom terminology. Either CSV or TMX.

TerminologyDataLocation

data TerminologyDataLocation

The location of the custom terminology data.

See: newTerminologyDataLocation smart constructor.

Constructors

TerminologyDataLocation'

Fields

repositoryType :: Text

The repository type for the custom terminology data.

location :: Text

The location of the custom terminology data.

newTerminologyDataLocation

Arguments

:: Text
$sel:repositoryType:TerminologyDataLocation'

-> Text
$sel:location:TerminologyDataLocation'

-> TerminologyDataLocation

Create a value of TerminologyDataLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositoryType:TerminologyDataLocation', terminologyDataLocation_repositoryType - The repository type for the custom terminology data.

$sel:location:TerminologyDataLocation', terminologyDataLocation_location - The location of the custom terminology data.

terminologyDataLocation_repositoryType :: Lens' TerminologyDataLocation Text

The repository type for the custom terminology data.

terminologyDataLocation_location :: Lens' TerminologyDataLocation Text

The location of the custom terminology data.

TerminologyProperties

data TerminologyProperties

The properties of the custom terminology.

See: newTerminologyProperties smart constructor.

Constructors

TerminologyProperties'

Fields

encryptionKey :: Maybe EncryptionKey

The encryption key for the custom terminology.

createdAt :: Maybe POSIX

The time at which the custom terminology was created, based on the timestamp.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the custom terminology.

targetLanguageCodes :: Maybe [Text]

The language codes for the target languages available with the custom terminology file. All possible target languages are returned in array.

name :: Maybe Text

The name of the custom terminology.

sizeBytes :: Maybe Int

The size of the file used when importing a custom terminology.

description :: Maybe Text

The description of the custom terminology properties.

termCount :: Maybe Int

The number of terms included in the custom terminology.

sourceLanguageCode :: Maybe Text

The language code for the source text of the translation request for which the custom terminology is being used.

lastUpdatedAt :: Maybe POSIX

The time at which the custom terminology was last update, based on the timestamp.

newTerminologyProperties :: TerminologyProperties

Create a value of TerminologyProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionKey:TerminologyProperties', terminologyProperties_encryptionKey - The encryption key for the custom terminology.

$sel:createdAt:TerminologyProperties', terminologyProperties_createdAt - The time at which the custom terminology was created, based on the timestamp.

$sel:arn:TerminologyProperties', terminologyProperties_arn - The Amazon Resource Name (ARN) of the custom terminology.

$sel:targetLanguageCodes:TerminologyProperties', terminologyProperties_targetLanguageCodes - The language codes for the target languages available with the custom terminology file. All possible target languages are returned in array.

$sel:name:TerminologyProperties', terminologyProperties_name - The name of the custom terminology.

$sel:sizeBytes:TerminologyProperties', terminologyProperties_sizeBytes - The size of the file used when importing a custom terminology.

$sel:description:TerminologyProperties', terminologyProperties_description - The description of the custom terminology properties.

$sel:termCount:TerminologyProperties', terminologyProperties_termCount - The number of terms included in the custom terminology.

$sel:sourceLanguageCode:TerminologyProperties', terminologyProperties_sourceLanguageCode - The language code for the source text of the translation request for which the custom terminology is being used.

$sel:lastUpdatedAt:TerminologyProperties', terminologyProperties_lastUpdatedAt - The time at which the custom terminology was last update, based on the timestamp.

terminologyProperties_encryptionKey :: Lens' TerminologyProperties (Maybe EncryptionKey)

The encryption key for the custom terminology.

terminologyProperties_createdAt :: Lens' TerminologyProperties (Maybe UTCTime)

The time at which the custom terminology was created, based on the timestamp.

terminologyProperties_arn :: Lens' TerminologyProperties (Maybe Text)

The Amazon Resource Name (ARN) of the custom terminology.

terminologyProperties_targetLanguageCodes :: Lens' TerminologyProperties (Maybe [Text])

The language codes for the target languages available with the custom terminology file. All possible target languages are returned in array.

terminologyProperties_name :: Lens' TerminologyProperties (Maybe Text)

The name of the custom terminology.

terminologyProperties_sizeBytes :: Lens' TerminologyProperties (Maybe Int)

The size of the file used when importing a custom terminology.

terminologyProperties_description :: Lens' TerminologyProperties (Maybe Text)

The description of the custom terminology properties.

terminologyProperties_termCount :: Lens' TerminologyProperties (Maybe Int)

The number of terms included in the custom terminology.

terminologyProperties_sourceLanguageCode :: Lens' TerminologyProperties (Maybe Text)

The language code for the source text of the translation request for which the custom terminology is being used.

terminologyProperties_lastUpdatedAt :: Lens' TerminologyProperties (Maybe UTCTime)

The time at which the custom terminology was last update, based on the timestamp.

TextTranslationJobFilter

data TextTranslationJobFilter

Provides information for filtering a list of translation jobs. For more information, see ListTextTranslationJobs.

See: newTextTranslationJobFilter smart constructor.

Constructors

TextTranslationJobFilter'

Fields

jobStatus :: Maybe JobStatus

Filters the list of jobs based by job status.

submittedAfterTime :: Maybe POSIX

Filters the list of jobs based on the time that the job was submitted for processing and returns only the jobs submitted after the specified time. Jobs are returned in descending order, newest to oldest.

submittedBeforeTime :: Maybe POSIX

Filters the list of jobs based on the time that the job was submitted for processing and returns only the jobs submitted before the specified time. Jobs are returned in ascending order, oldest to newest.

jobName :: Maybe Text

Filters the list of jobs by name.

newTextTranslationJobFilter :: TextTranslationJobFilter

Create a value of TextTranslationJobFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobStatus:TextTranslationJobFilter', textTranslationJobFilter_jobStatus - Filters the list of jobs based by job status.

$sel:submittedAfterTime:TextTranslationJobFilter', textTranslationJobFilter_submittedAfterTime - Filters the list of jobs based on the time that the job was submitted for processing and returns only the jobs submitted after the specified time. Jobs are returned in descending order, newest to oldest.

$sel:submittedBeforeTime:TextTranslationJobFilter', textTranslationJobFilter_submittedBeforeTime - Filters the list of jobs based on the time that the job was submitted for processing and returns only the jobs submitted before the specified time. Jobs are returned in ascending order, oldest to newest.

$sel:jobName:TextTranslationJobFilter', textTranslationJobFilter_jobName - Filters the list of jobs by name.

textTranslationJobFilter_jobStatus :: Lens' TextTranslationJobFilter (Maybe JobStatus)

Filters the list of jobs based by job status.

textTranslationJobFilter_submittedAfterTime :: Lens' TextTranslationJobFilter (Maybe UTCTime)

Filters the list of jobs based on the time that the job was submitted for processing and returns only the jobs submitted after the specified time. Jobs are returned in descending order, newest to oldest.

textTranslationJobFilter_submittedBeforeTime :: Lens' TextTranslationJobFilter (Maybe UTCTime)

Filters the list of jobs based on the time that the job was submitted for processing and returns only the jobs submitted before the specified time. Jobs are returned in ascending order, oldest to newest.

textTranslationJobFilter_jobName :: Lens' TextTranslationJobFilter (Maybe Text)

Filters the list of jobs by name.

TextTranslationJobProperties

data TextTranslationJobProperties

Provides information about a translation job.

See: newTextTranslationJobProperties smart constructor.

Constructors

TextTranslationJobProperties'

Fields

parallelDataNames :: Maybe [Text]

A list containing the names of the parallel data resources applied to the translation job.

submittedTime :: Maybe POSIX

The time at which the translation job was submitted.

inputDataConfig :: Maybe InputDataConfig

The input configuration properties that were specified when the job was requested.

message :: Maybe Text

An explanation of any errors that may have occured during the translation job.

jobStatus :: Maybe JobStatus

The status of the translation job.

outputDataConfig :: Maybe OutputDataConfig

The output configuration properties that were specified when the job was requested.

jobDetails :: Maybe JobDetails

The number of documents successfully and unsuccessfully processed during the translation job.

targetLanguageCodes :: Maybe (NonEmpty Text)

The language code of the language of the target text. The language must be a language supported by Amazon Translate.

endTime :: Maybe POSIX

The time at which the translation job ended.

terminologyNames :: Maybe [Text]

A list containing the names of the terminologies applied to a translation job. Only one terminology can be applied per StartTextTranslationJob request at this time.

jobName :: Maybe Text

The user-defined name of the translation job.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of an AWS Identity Access and Management (IAM) role that granted Amazon Translate read access to the job's input data.

sourceLanguageCode :: Maybe Text

The language code of the language of the source text. The language must be a language supported by Amazon Translate.

jobId :: Maybe Text

The ID of the translation job.

newTextTranslationJobProperties :: TextTranslationJobProperties

Create a value of TextTranslationJobProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parallelDataNames:TextTranslationJobProperties', textTranslationJobProperties_parallelDataNames - A list containing the names of the parallel data resources applied to the translation job.

$sel:submittedTime:TextTranslationJobProperties', textTranslationJobProperties_submittedTime - The time at which the translation job was submitted.

$sel:inputDataConfig:TextTranslationJobProperties', textTranslationJobProperties_inputDataConfig - The input configuration properties that were specified when the job was requested.

$sel:message:TextTranslationJobProperties', textTranslationJobProperties_message - An explanation of any errors that may have occured during the translation job.

$sel:jobStatus:TextTranslationJobProperties', textTranslationJobProperties_jobStatus - The status of the translation job.

$sel:outputDataConfig:TextTranslationJobProperties', textTranslationJobProperties_outputDataConfig - The output configuration properties that were specified when the job was requested.

$sel:jobDetails:TextTranslationJobProperties', textTranslationJobProperties_jobDetails - The number of documents successfully and unsuccessfully processed during the translation job.

$sel:targetLanguageCodes:TextTranslationJobProperties', textTranslationJobProperties_targetLanguageCodes - The language code of the language of the target text. The language must be a language supported by Amazon Translate.

$sel:endTime:TextTranslationJobProperties', textTranslationJobProperties_endTime - The time at which the translation job ended.

$sel:terminologyNames:TextTranslationJobProperties', textTranslationJobProperties_terminologyNames - A list containing the names of the terminologies applied to a translation job. Only one terminology can be applied per StartTextTranslationJob request at this time.

$sel:jobName:TextTranslationJobProperties', textTranslationJobProperties_jobName - The user-defined name of the translation job.

$sel:dataAccessRoleArn:TextTranslationJobProperties', textTranslationJobProperties_dataAccessRoleArn - The Amazon Resource Name (ARN) of an AWS Identity Access and Management (IAM) role that granted Amazon Translate read access to the job's input data.

$sel:sourceLanguageCode:TextTranslationJobProperties', textTranslationJobProperties_sourceLanguageCode - The language code of the language of the source text. The language must be a language supported by Amazon Translate.

$sel:jobId:TextTranslationJobProperties', textTranslationJobProperties_jobId - The ID of the translation job.

textTranslationJobProperties_parallelDataNames :: Lens' TextTranslationJobProperties (Maybe [Text])

A list containing the names of the parallel data resources applied to the translation job.

textTranslationJobProperties_submittedTime :: Lens' TextTranslationJobProperties (Maybe UTCTime)

The time at which the translation job was submitted.

textTranslationJobProperties_inputDataConfig :: Lens' TextTranslationJobProperties (Maybe InputDataConfig)

The input configuration properties that were specified when the job was requested.

textTranslationJobProperties_message :: Lens' TextTranslationJobProperties (Maybe Text)

An explanation of any errors that may have occured during the translation job.

textTranslationJobProperties_jobStatus :: Lens' TextTranslationJobProperties (Maybe JobStatus)

The status of the translation job.

textTranslationJobProperties_outputDataConfig :: Lens' TextTranslationJobProperties (Maybe OutputDataConfig)

The output configuration properties that were specified when the job was requested.

textTranslationJobProperties_jobDetails :: Lens' TextTranslationJobProperties (Maybe JobDetails)

The number of documents successfully and unsuccessfully processed during the translation job.

textTranslationJobProperties_targetLanguageCodes :: Lens' TextTranslationJobProperties (Maybe (NonEmpty Text))

The language code of the language of the target text. The language must be a language supported by Amazon Translate.

textTranslationJobProperties_endTime :: Lens' TextTranslationJobProperties (Maybe UTCTime)

The time at which the translation job ended.

textTranslationJobProperties_terminologyNames :: Lens' TextTranslationJobProperties (Maybe [Text])

A list containing the names of the terminologies applied to a translation job. Only one terminology can be applied per StartTextTranslationJob request at this time.

textTranslationJobProperties_jobName :: Lens' TextTranslationJobProperties (Maybe Text)

The user-defined name of the translation job.

textTranslationJobProperties_dataAccessRoleArn :: Lens' TextTranslationJobProperties (Maybe Text)

The Amazon Resource Name (ARN) of an AWS Identity Access and Management (IAM) role that granted Amazon Translate read access to the job's input data.

textTranslationJobProperties_sourceLanguageCode :: Lens' TextTranslationJobProperties (Maybe Text)

The language code of the language of the source text. The language must be a language supported by Amazon Translate.

textTranslationJobProperties_jobId :: Lens' TextTranslationJobProperties (Maybe Text)

The ID of the translation job.