Network.AWS.Translate.Types.TerminologyData

Description

Documentation

data TerminologyData

The data associated with the custom terminology.

See: newTerminologyData smart constructor.

Constructors

TerminologyData'

Fields

file :: Sensitive Base64

The file containing the custom terminology data. Your version of the AWS SDK performs a Base64-encoding on this field before sending a request to the AWS service. Users of the SDK should not perform Base64-encoding themselves.

format :: TerminologyDataFormat

The data format of the custom terminology. Either CSV or TMX.

newTerminologyData

Arguments

:: ByteString
$sel:file:TerminologyData'

-> TerminologyDataFormat
$sel:format:TerminologyData'

-> TerminologyData

Create a value of TerminologyData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:file:TerminologyData', terminologyData_file - The file containing the custom terminology data. Your version of the AWS SDK performs a Base64-encoding on this field before sending a request to the AWS service. Users of the SDK should not perform Base64-encoding themselves.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:format:TerminologyData', terminologyData_format - The data format of the custom terminology. Either CSV or TMX.

terminologyData_file :: Lens' TerminologyData ByteString

The file containing the custom terminology data. Your version of the AWS SDK performs a Base64-encoding on this field before sending a request to the AWS service. Users of the SDK should not perform Base64-encoding themselves.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

terminologyData_format :: Lens' TerminologyData TerminologyDataFormat

The data format of the custom terminology. Either CSV or TMX.