Network.AWS.Translate.Types.TextTranslationJobProperties

Description

Documentation

data TextTranslationJobProperties

Provides information about a translation job.

See: newTextTranslationJobProperties smart constructor.

Constructors

TextTranslationJobProperties'

Fields

parallelDataNames :: Maybe [Text]

A list containing the names of the parallel data resources applied to the translation job.

submittedTime :: Maybe POSIX

The time at which the translation job was submitted.

inputDataConfig :: Maybe InputDataConfig

The input configuration properties that were specified when the job was requested.

message :: Maybe Text

An explanation of any errors that may have occured during the translation job.

jobStatus :: Maybe JobStatus

The status of the translation job.

outputDataConfig :: Maybe OutputDataConfig

The output configuration properties that were specified when the job was requested.

jobDetails :: Maybe JobDetails

The number of documents successfully and unsuccessfully processed during the translation job.

targetLanguageCodes :: Maybe (NonEmpty Text)

The language code of the language of the target text. The language must be a language supported by Amazon Translate.

endTime :: Maybe POSIX

The time at which the translation job ended.

terminologyNames :: Maybe [Text]

A list containing the names of the terminologies applied to a translation job. Only one terminology can be applied per StartTextTranslationJob request at this time.

jobName :: Maybe Text

The user-defined name of the translation job.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of an AWS Identity Access and Management (IAM) role that granted Amazon Translate read access to the job's input data.

sourceLanguageCode :: Maybe Text

The language code of the language of the source text. The language must be a language supported by Amazon Translate.

jobId :: Maybe Text

The ID of the translation job.

newTextTranslationJobProperties :: TextTranslationJobProperties

Create a value of TextTranslationJobProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parallelDataNames:TextTranslationJobProperties', textTranslationJobProperties_parallelDataNames - A list containing the names of the parallel data resources applied to the translation job.

$sel:submittedTime:TextTranslationJobProperties', textTranslationJobProperties_submittedTime - The time at which the translation job was submitted.

$sel:inputDataConfig:TextTranslationJobProperties', textTranslationJobProperties_inputDataConfig - The input configuration properties that were specified when the job was requested.

$sel:message:TextTranslationJobProperties', textTranslationJobProperties_message - An explanation of any errors that may have occured during the translation job.

$sel:jobStatus:TextTranslationJobProperties', textTranslationJobProperties_jobStatus - The status of the translation job.

$sel:outputDataConfig:TextTranslationJobProperties', textTranslationJobProperties_outputDataConfig - The output configuration properties that were specified when the job was requested.

$sel:jobDetails:TextTranslationJobProperties', textTranslationJobProperties_jobDetails - The number of documents successfully and unsuccessfully processed during the translation job.

$sel:targetLanguageCodes:TextTranslationJobProperties', textTranslationJobProperties_targetLanguageCodes - The language code of the language of the target text. The language must be a language supported by Amazon Translate.

$sel:endTime:TextTranslationJobProperties', textTranslationJobProperties_endTime - The time at which the translation job ended.

$sel:terminologyNames:TextTranslationJobProperties', textTranslationJobProperties_terminologyNames - A list containing the names of the terminologies applied to a translation job. Only one terminology can be applied per StartTextTranslationJob request at this time.

$sel:jobName:TextTranslationJobProperties', textTranslationJobProperties_jobName - The user-defined name of the translation job.

$sel:dataAccessRoleArn:TextTranslationJobProperties', textTranslationJobProperties_dataAccessRoleArn - The Amazon Resource Name (ARN) of an AWS Identity Access and Management (IAM) role that granted Amazon Translate read access to the job's input data.

$sel:sourceLanguageCode:TextTranslationJobProperties', textTranslationJobProperties_sourceLanguageCode - The language code of the language of the source text. The language must be a language supported by Amazon Translate.

$sel:jobId:TextTranslationJobProperties', textTranslationJobProperties_jobId - The ID of the translation job.

textTranslationJobProperties_parallelDataNames :: Lens' TextTranslationJobProperties (Maybe [Text])

A list containing the names of the parallel data resources applied to the translation job.

textTranslationJobProperties_submittedTime :: Lens' TextTranslationJobProperties (Maybe UTCTime)

The time at which the translation job was submitted.

textTranslationJobProperties_inputDataConfig :: Lens' TextTranslationJobProperties (Maybe InputDataConfig)

The input configuration properties that were specified when the job was requested.

textTranslationJobProperties_message :: Lens' TextTranslationJobProperties (Maybe Text)

An explanation of any errors that may have occured during the translation job.

textTranslationJobProperties_jobStatus :: Lens' TextTranslationJobProperties (Maybe JobStatus)

The status of the translation job.

textTranslationJobProperties_outputDataConfig :: Lens' TextTranslationJobProperties (Maybe OutputDataConfig)

The output configuration properties that were specified when the job was requested.

textTranslationJobProperties_jobDetails :: Lens' TextTranslationJobProperties (Maybe JobDetails)

The number of documents successfully and unsuccessfully processed during the translation job.

textTranslationJobProperties_targetLanguageCodes :: Lens' TextTranslationJobProperties (Maybe (NonEmpty Text))

The language code of the language of the target text. The language must be a language supported by Amazon Translate.

textTranslationJobProperties_endTime :: Lens' TextTranslationJobProperties (Maybe UTCTime)

The time at which the translation job ended.

textTranslationJobProperties_terminologyNames :: Lens' TextTranslationJobProperties (Maybe [Text])

A list containing the names of the terminologies applied to a translation job. Only one terminology can be applied per StartTextTranslationJob request at this time.

textTranslationJobProperties_jobName :: Lens' TextTranslationJobProperties (Maybe Text)

The user-defined name of the translation job.

textTranslationJobProperties_dataAccessRoleArn :: Lens' TextTranslationJobProperties (Maybe Text)

The Amazon Resource Name (ARN) of an AWS Identity Access and Management (IAM) role that granted Amazon Translate read access to the job's input data.

textTranslationJobProperties_sourceLanguageCode :: Lens' TextTranslationJobProperties (Maybe Text)

The language code of the language of the source text. The language must be a language supported by Amazon Translate.

textTranslationJobProperties_jobId :: Lens' TextTranslationJobProperties (Maybe Text)

The ID of the translation job.