Network.AWS.Types

Description

Authentication

Credentials

newtype AccessKey

An access key ID.

For example: AKIAIOSFODNN7EXAMPLE

See: Understanding and Getting Your Security Credentials.

Constructors

AccessKey ByteString

newtype SecretKey

Secret access key credential.

For example: wJalrXUtnFEMIK7MDENGbPxRfiCYEXAMPLEKE

See: Understanding and Getting Your Security Credentials.

Constructors

SecretKey ByteString

newtype SessionToken

A session token used by STS to temporarily authorise access to an AWS resource.

See: Temporary Security Credentials.

Constructors

SessionToken ByteString

Environment

data Auth

An authorisation environment containing AWS credentials, and potentially a reference which can be refreshed out-of-band as temporary credentials expire.

Constructors

Ref ThreadId (IORef AuthEnv)

Auth AuthEnv

withAuth :: MonadIO m => Auth -> (AuthEnv -> m a) -> m a

data AuthEnv

The AuthN/AuthZ credential environment.

Constructors

AuthEnv

Fields

_authAccessKeyId :: AccessKey

_authSecretAccessKey :: Sensitive SecretKey

_authSessionToken :: Maybe (Sensitive SessionToken)

_authExpiration :: Maybe ISO8601

authAccessKeyId :: Lens' AuthEnv AccessKey

The access key ID that identifies the temporary security credentials.

authSecretAccessKey :: Lens' AuthEnv SecretKey

The secret access key that can be used to sign requests.

authSessionToken :: Lens' AuthEnv (Maybe SessionToken)

The token that users must pass to the service API to use the temporary credentials.

authExpiration :: Lens' AuthEnv (Maybe UTCTime)

The date on which the current credentials expire.

Logging

data LogLevel

Constructors

Info
Info messages supplied by the user - this level is not emitted by the library.

Error
Error messages only.

Debug
Useful debug information + info + error levels.

Trace
Includes potentially sensitive signing metadata, and non-streaming response bodies.

type Logger = LogLevel -> ByteStringBuilder -> IO ()

A function threaded through various request and serialisation routines to log informational and debug messages.

Signing

type Algorithm a = Request a -> AuthEnv -> Region -> UTCTime -> Signed a

data Meta where

Signing algorithm specific metadata.

Constructors

Meta :: ToLog a => a -> Meta

data Signer

Constructors

Signer

Fields

signerSign :: forall a. Algorithm a

signerPresign :: forall a. Seconds -> Algorithm a

data Signed a

A signed ClientRequest and associated metadata specific to the signing algorithm, tagged with the initial request type to be able to obtain the associated response, 'Rs a'.

Constructors

Signed

Fields

signedMeta :: Meta

signedRequest :: ClientRequest

Service

data Abbrev

Abbreviated service name.

data Service

Attributes and functions specific to an AWS service.

Constructors

Service

Fields

_serviceAbbrev :: Abbrev

_serviceSigner :: Signer

_serviceSigningName :: ByteString

_serviceVersion :: ByteString

_serviceEndpointPrefix :: ByteString

_serviceEndpoint :: Region -> Endpoint

_serviceTimeout :: Maybe Seconds

_serviceCheck :: Status -> Bool

_serviceError :: Status -> [Header] -> ByteStringLazy -> Error

_serviceRetry :: Retry

serviceSigner :: Lens' Service Signer

serviceEndpoint :: Setter' Service Endpoint

serviceTimeout :: Lens' Service (Maybe Seconds)

serviceCheck :: Lens' Service (Status -> Bool)

serviceRetry :: Lens' Service Retry

Requests

class AWSRequest a where

Specify how a request can be de/serialised.

Associated Types

type AWSResponse a :: *

The successful, expected response associated with a request.

Methods

request :: a -> Request a

response :: MonadResource m => Logger -> Service -> Proxy a -> ClientResponse ClientBody -> m (Either Error (ClientResponse (AWSResponse a)))

data Request a

An unsigned request.

Constructors

Request

Fields

_requestService :: Service

_requestMethod :: StdMethod

_requestPath :: RawPath

_requestQuery :: QueryString

_requestHeaders :: [Header]

_requestBody :: RequestBody

requestService :: Lens' (Request a) Service

requestMethod :: Lens' (Request a) StdMethod

requestHeaders :: Lens' (Request a) [Header]

requestPath :: Lens' (Request a) RawPath

requestQuery :: Lens' (Request a) QueryString

requestBody :: Lens' (Request a) RequestBody

requestSign :: Algorithm a

requestPresign :: Seconds -> Algorithm a

Retries

data Retry

Constants and predicates used to create a RetryPolicy.

Constructors

Exponential

Fields

_retryBase :: Double

_retryGrowth :: Int

_retryAttempts :: Int

_retryCheck :: ServiceError -> Maybe Text

Returns a descriptive name for logging if the request should be retried.

exponentBase :: Lens' Retry Double

exponentGrowth :: Lens' Retry Int

retryAttempts :: Lens' Retry Int

retryCheck :: Lens' Retry (ServiceError -> Maybe Text)

Errors

class AsError a where

Minimal complete definition

_Error

Methods

_Error :: Prism' a Error

A general Amazonka error.

_TransportError :: Prism' a HttpException

An error occured while communicating over HTTP with a remote service.

_SerializeError :: Prism' a SerializeError

A serialisation error occured when attempting to deserialise a response.

_ServiceError :: Prism' a ServiceError

A service specific error returned by the remote service.

data Error

An error type representing errors that can be attributed to this library.

Constructors

TransportError HttpException

SerializeError SerializeError

ServiceError ServiceError

HTTP Errors

data HttpException

An exception which may be generated by this library

Since: http-client-0.5.0

Serialize Errors

data SerializeError

Constructors

SerializeError'

Fields

_serializeErrorAbbrev :: Abbrev

_serializeErrorStatus :: Status

_serializeErrorBody :: Maybe ByteStringLazy

The response body, if the response was not streaming.

_serializeErrorMessage :: String

serializeAbbrev :: Lens' SerializeError Abbrev

serializeStatus :: Lens' SerializeError Status

serializeMessage :: Lens' SerializeError String

Service Errors

data ServiceError

Constructors

ServiceError'

Fields

_serviceErrorAbbrev :: Abbrev

_serviceErrorStatus :: Status

_serviceErrorHeaders :: [Header]

_serviceErrorCode :: ErrorCode

_serviceErrorMessage :: Maybe ErrorMessage

_serviceErrorRequestId :: Maybe RequestId

serviceAbbrev :: Lens' ServiceError Abbrev

serviceStatus :: Lens' ServiceError Status

serviceHeaders :: Lens' ServiceError [Header]

serviceCode :: Lens' ServiceError ErrorCode

serviceMessage :: Lens' ServiceError (Maybe ErrorMessage)

serviceRequestId :: Lens' ServiceError (Maybe RequestId)

Error Types

newtype ErrorCode

Constructors

ErrorCode Text

newErrorCode :: Text -> ErrorCode

Construct an ErrorCode.

newtype ErrorMessage

Constructors

ErrorMessage

Fields

fromErrorMessage :: Text

newtype RequestId

Constructors

RequestId

Fields

fromRequestId :: Text

Regions

newtype Region

The available AWS regions.

Constructors

Region'

Fields

fromRegion :: Text

Bundled Patterns

pattern NorthVirginia :: Region

pattern Ohio :: Region

pattern NorthCalifornia :: Region

pattern Oregon :: Region

pattern GovCloudWest :: Region

pattern GovCloudEast :: Region

pattern Montreal :: Region

pattern SaoPaulo :: Region

pattern Frankfurt :: Region

pattern Ireland :: Region

pattern London :: Region

pattern Milan :: Region

pattern Paris :: Region

pattern Stockholm :: Region

pattern Bahrain :: Region

pattern CapeTown :: Region

pattern Beijing :: Region

pattern Ningxia :: Region

pattern HongKong :: Region

pattern Tokyo :: Region

pattern Seoul :: Region

pattern Osaka :: Region

pattern Singapore :: Region

pattern Sydney :: Region

pattern Mumbai :: Region

Endpoints

data Endpoint

Constructors

Endpoint

Fields

_endpointHost :: ByteString

_endpointSecure :: Bool

_endpointPort :: Int

_endpointScope :: ByteString

endpointHost :: Lens' Endpoint ByteString

endpointPort :: Lens' Endpoint Int

endpointSecure :: Lens' Endpoint Bool

endpointScope :: Lens' Endpoint ByteString

HTTP

type ClientRequest = Request

A convenience alias to avoid type ambiguity.

type ClientResponse = Response

A convenience alias encapsulating the common Response.

type ClientBody = ConduitM () ByteString (ResourceT IO) ()

A convenience alias encapsulating the common Response body.

newClientRequest :: Endpoint -> Maybe Seconds -> ClientRequest

Construct a ClientRequest using common parameters such as TLS and prevent throwing errors when receiving erroneous status codes in respones.

Seconds

newtype Seconds

An integral value representing seconds.

Constructors

Seconds Int

toSeconds :: Seconds -> Int

toMicroseconds :: Seconds -> Int