Network.AWS.VoiceId.EvaluateSession

Description

Evaluates a specified session based on audio data accumulated during a streaming Amazon Connect Voice ID call.

Creating a Request

data EvaluateSession

See: newEvaluateSession smart constructor.

Constructors

EvaluateSession'

Fields

domainId :: Text

The identifier of the domain where the session started.

sessionNameOrId :: Text

The session identifier, or name of the session, that you want to evaluate. In Voice ID integration, this is the Contact-Id.

newEvaluateSession

Arguments

:: Text
$sel:domainId:EvaluateSession'

-> Text
$sel:sessionNameOrId:EvaluateSession'

-> EvaluateSession

Create a value of EvaluateSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainId:EvaluateSession', evaluateSession_domainId - The identifier of the domain where the session started.

$sel:sessionNameOrId:EvaluateSession', evaluateSession_sessionNameOrId - The session identifier, or name of the session, that you want to evaluate. In Voice ID integration, this is the Contact-Id.

Request Lenses

evaluateSession_domainId :: Lens' EvaluateSession Text

The identifier of the domain where the session started.

evaluateSession_sessionNameOrId :: Lens' EvaluateSession Text

The session identifier, or name of the session, that you want to evaluate. In Voice ID integration, this is the Contact-Id.

Destructuring the Response

data EvaluateSessionResponse

See: newEvaluateSessionResponse smart constructor.

Constructors

EvaluateSessionResponse'

Fields

fraudDetectionResult :: Maybe FraudDetectionResult

Details resulting from the fraud detection process, such as fraud detection decision and risk score.

streamingStatus :: Maybe StreamingStatus

The current status of audio streaming for this session. This field is useful to infer next steps when the Authentication or Fraud Detection results are empty or the decision is NOT_ENOUGH_SPEECH. In this situation, if the StreamingStatus is ONGOING/PENDING_CONFIGURATION, it can mean that the client should call the API again later, once Voice ID has enough audio to produce a result. If the decision remains NOT_ENOUGH_SPEECH even after StreamingStatus is ENDED, it means that the previously streamed session did not have enough speech to perform evaluation, and a new streaming session is needed to try again.

authenticationResult :: Maybe AuthenticationResult

Details resulting from the authentication process, such as authentication decision and authentication score.

domainId :: Maybe Text

The identifier of the domain containing the session.

sessionId :: Maybe Text

The service-generated identifier of the session.

sessionName :: Maybe Text

The client-provided name of the session.

httpStatus :: Int

The response's http status code.

newEvaluateSessionResponse

Arguments

:: Int
$sel:httpStatus:EvaluateSessionResponse'

-> EvaluateSessionResponse

Create a value of EvaluateSessionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fraudDetectionResult:EvaluateSessionResponse', evaluateSessionResponse_fraudDetectionResult - Details resulting from the fraud detection process, such as fraud detection decision and risk score.

$sel:streamingStatus:EvaluateSessionResponse', evaluateSessionResponse_streamingStatus - The current status of audio streaming for this session. This field is useful to infer next steps when the Authentication or Fraud Detection results are empty or the decision is NOT_ENOUGH_SPEECH. In this situation, if the StreamingStatus is ONGOING/PENDING_CONFIGURATION, it can mean that the client should call the API again later, once Voice ID has enough audio to produce a result. If the decision remains NOT_ENOUGH_SPEECH even after StreamingStatus is ENDED, it means that the previously streamed session did not have enough speech to perform evaluation, and a new streaming session is needed to try again.

$sel:authenticationResult:EvaluateSessionResponse', evaluateSessionResponse_authenticationResult - Details resulting from the authentication process, such as authentication decision and authentication score.

$sel:domainId:EvaluateSession', evaluateSessionResponse_domainId - The identifier of the domain containing the session.

$sel:sessionId:EvaluateSessionResponse', evaluateSessionResponse_sessionId - The service-generated identifier of the session.

$sel:sessionName:EvaluateSessionResponse', evaluateSessionResponse_sessionName - The client-provided name of the session.

$sel:httpStatus:EvaluateSessionResponse', evaluateSessionResponse_httpStatus - The response's http status code.

Response Lenses

evaluateSessionResponse_fraudDetectionResult :: Lens' EvaluateSessionResponse (Maybe FraudDetectionResult)

Details resulting from the fraud detection process, such as fraud detection decision and risk score.

evaluateSessionResponse_streamingStatus :: Lens' EvaluateSessionResponse (Maybe StreamingStatus)

The current status of audio streaming for this session. This field is useful to infer next steps when the Authentication or Fraud Detection results are empty or the decision is NOT_ENOUGH_SPEECH. In this situation, if the StreamingStatus is ONGOING/PENDING_CONFIGURATION, it can mean that the client should call the API again later, once Voice ID has enough audio to produce a result. If the decision remains NOT_ENOUGH_SPEECH even after StreamingStatus is ENDED, it means that the previously streamed session did not have enough speech to perform evaluation, and a new streaming session is needed to try again.

evaluateSessionResponse_authenticationResult :: Lens' EvaluateSessionResponse (Maybe AuthenticationResult)

Details resulting from the authentication process, such as authentication decision and authentication score.

evaluateSessionResponse_domainId :: Lens' EvaluateSessionResponse (Maybe Text)

The identifier of the domain containing the session.

evaluateSessionResponse_sessionId :: Lens' EvaluateSessionResponse (Maybe Text)

The service-generated identifier of the session.

evaluateSessionResponse_sessionName :: Lens' EvaluateSessionResponse (Maybe Text)

The client-provided name of the session.

evaluateSessionResponse_httpStatus :: Lens' EvaluateSessionResponse Int

The response's http status code.