Network.AWS.VoiceId.ListSpeakers

Description

Lists all speakers in a specified domain.

Creating a Request

data ListSpeakers

See: newListSpeakers smart constructor.

Constructors

ListSpeakers'

Fields

nextToken :: Maybe Text

If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

maxResults :: Maybe Natural

The maximum number of results that are returned per call. You can use NextToken to obtain further pages of results. The default is 100; the maximum allowed page size is also 100.

domainId :: Text

The identifier of the domain.

newListSpeakers

Arguments

:: Text
$sel:domainId:ListSpeakers'

-> ListSpeakers

Create a value of ListSpeakers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSpeakers', listSpeakers_nextToken - If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

$sel:maxResults:ListSpeakers', listSpeakers_maxResults - The maximum number of results that are returned per call. You can use NextToken to obtain further pages of results. The default is 100; the maximum allowed page size is also 100.

$sel:domainId:ListSpeakers', listSpeakers_domainId - The identifier of the domain.

Request Lenses

listSpeakers_nextToken :: Lens' ListSpeakers (Maybe Text)

If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

listSpeakers_maxResults :: Lens' ListSpeakers (Maybe Natural)

The maximum number of results that are returned per call. You can use NextToken to obtain further pages of results. The default is 100; the maximum allowed page size is also 100.

listSpeakers_domainId :: Lens' ListSpeakers Text

The identifier of the domain.

Destructuring the Response

data ListSpeakersResponse

See: newListSpeakersResponse smart constructor.

Constructors

ListSpeakersResponse'

Fields

speakerSummaries :: Maybe [SpeakerSummary]

A list containing details about each speaker in the Amazon Web Services account.

nextToken :: Maybe Text

If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

httpStatus :: Int

The response's http status code.

newListSpeakersResponse

Arguments

:: Int
$sel:httpStatus:ListSpeakersResponse'

-> ListSpeakersResponse

Create a value of ListSpeakersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:speakerSummaries:ListSpeakersResponse', listSpeakersResponse_speakerSummaries - A list containing details about each speaker in the Amazon Web Services account.

$sel:nextToken:ListSpeakers', listSpeakersResponse_nextToken - If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

$sel:httpStatus:ListSpeakersResponse', listSpeakersResponse_httpStatus - The response's http status code.

Response Lenses

listSpeakersResponse_speakerSummaries :: Lens' ListSpeakersResponse (Maybe [SpeakerSummary])

A list containing details about each speaker in the Amazon Web Services account.

listSpeakersResponse_nextToken :: Lens' ListSpeakersResponse (Maybe Text)

If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

listSpeakersResponse_httpStatus :: Lens' ListSpeakersResponse Int

The response's http status code.