Network.AWS.VoiceId.StartFraudsterRegistrationJob

Description

Starts a new batch fraudster registration job using provided details.

Creating a Request

data StartFraudsterRegistrationJob

See: newStartFraudsterRegistrationJob smart constructor.

Constructors

StartFraudsterRegistrationJob'

Fields

clientToken :: Maybe Text

The idempotency token for starting a new fraudster registration job. If not provided, Amazon Web Services SDK populates this field.

jobName :: Maybe (Sensitive Text)

The name of the new fraudster registration job.

registrationConfig :: Maybe RegistrationConfig

The registration config containing details such as the action to take when a duplicate fraudster is detected, and the similarity threshold to use for detecting a duplicate fraudster.

dataAccessRoleArn :: Text

The IAM role Amazon Resource Name (ARN) that grants Voice ID permissions to access customer's buckets to read the input manifest file and write the Job output file. Refer to the Create and edit a fraudster watchlist documentation for the permissions needed in this role.

domainId :: Text

The identifier of the domain containing the fraudster registration job and in which the fraudsters are registered.

inputDataConfig :: InputDataConfig

The input data config containing an S3 URI for the input manifest file that contains the list of fraudster registration requests.

outputDataConfig :: OutputDataConfig

The output data config containing the S3 location where Voice ID writes the job output file; you must also include a KMS Key ID to encrypt the file.

newStartFraudsterRegistrationJob

Arguments

:: Text
$sel:dataAccessRoleArn:StartFraudsterRegistrationJob'

-> Text
$sel:domainId:StartFraudsterRegistrationJob'

-> InputDataConfig
$sel:inputDataConfig:StartFraudsterRegistrationJob'

-> OutputDataConfig
$sel:outputDataConfig:StartFraudsterRegistrationJob'

-> StartFraudsterRegistrationJob

Create a value of StartFraudsterRegistrationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:StartFraudsterRegistrationJob', startFraudsterRegistrationJob_clientToken - The idempotency token for starting a new fraudster registration job. If not provided, Amazon Web Services SDK populates this field.

$sel:jobName:StartFraudsterRegistrationJob', startFraudsterRegistrationJob_jobName - The name of the new fraudster registration job.

$sel:registrationConfig:StartFraudsterRegistrationJob', startFraudsterRegistrationJob_registrationConfig - The registration config containing details such as the action to take when a duplicate fraudster is detected, and the similarity threshold to use for detecting a duplicate fraudster.

$sel:dataAccessRoleArn:StartFraudsterRegistrationJob', startFraudsterRegistrationJob_dataAccessRoleArn - The IAM role Amazon Resource Name (ARN) that grants Voice ID permissions to access customer's buckets to read the input manifest file and write the Job output file. Refer to the Create and edit a fraudster watchlist documentation for the permissions needed in this role.

$sel:domainId:StartFraudsterRegistrationJob', startFraudsterRegistrationJob_domainId - The identifier of the domain containing the fraudster registration job and in which the fraudsters are registered.

$sel:inputDataConfig:StartFraudsterRegistrationJob', startFraudsterRegistrationJob_inputDataConfig - The input data config containing an S3 URI for the input manifest file that contains the list of fraudster registration requests.

$sel:outputDataConfig:StartFraudsterRegistrationJob', startFraudsterRegistrationJob_outputDataConfig - The output data config containing the S3 location where Voice ID writes the job output file; you must also include a KMS Key ID to encrypt the file.

Request Lenses

startFraudsterRegistrationJob_clientToken :: Lens' StartFraudsterRegistrationJob (Maybe Text)

The idempotency token for starting a new fraudster registration job. If not provided, Amazon Web Services SDK populates this field.

startFraudsterRegistrationJob_jobName :: Lens' StartFraudsterRegistrationJob (Maybe Text)

The name of the new fraudster registration job.

startFraudsterRegistrationJob_registrationConfig :: Lens' StartFraudsterRegistrationJob (Maybe RegistrationConfig)

The registration config containing details such as the action to take when a duplicate fraudster is detected, and the similarity threshold to use for detecting a duplicate fraudster.

startFraudsterRegistrationJob_dataAccessRoleArn :: Lens' StartFraudsterRegistrationJob Text

The IAM role Amazon Resource Name (ARN) that grants Voice ID permissions to access customer's buckets to read the input manifest file and write the Job output file. Refer to the Create and edit a fraudster watchlist documentation for the permissions needed in this role.

startFraudsterRegistrationJob_domainId :: Lens' StartFraudsterRegistrationJob Text

The identifier of the domain containing the fraudster registration job and in which the fraudsters are registered.

startFraudsterRegistrationJob_inputDataConfig :: Lens' StartFraudsterRegistrationJob InputDataConfig

The input data config containing an S3 URI for the input manifest file that contains the list of fraudster registration requests.

startFraudsterRegistrationJob_outputDataConfig :: Lens' StartFraudsterRegistrationJob OutputDataConfig

The output data config containing the S3 location where Voice ID writes the job output file; you must also include a KMS Key ID to encrypt the file.

Destructuring the Response

data StartFraudsterRegistrationJobResponse

See: newStartFraudsterRegistrationJobResponse smart constructor.

Constructors

StartFraudsterRegistrationJobResponse'

Fields

job :: Maybe FraudsterRegistrationJob

Details about the started fraudster registration job.

httpStatus :: Int

The response's http status code.

newStartFraudsterRegistrationJobResponse

Arguments

:: Int
$sel:httpStatus:StartFraudsterRegistrationJobResponse'

-> StartFraudsterRegistrationJobResponse

Create a value of StartFraudsterRegistrationJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:job:StartFraudsterRegistrationJobResponse', startFraudsterRegistrationJobResponse_job - Details about the started fraudster registration job.

$sel:httpStatus:StartFraudsterRegistrationJobResponse', startFraudsterRegistrationJobResponse_httpStatus - The response's http status code.

Response Lenses

startFraudsterRegistrationJobResponse_job :: Lens' StartFraudsterRegistrationJobResponse (Maybe FraudsterRegistrationJob)

Details about the started fraudster registration job.

startFraudsterRegistrationJobResponse_httpStatus :: Lens' StartFraudsterRegistrationJobResponse Int

The response's http status code.