Network.AWS.VoiceId.StartSpeakerEnrollmentJob

Description

Starts a new batch speaker enrollment job using specified details.

Creating a Request

data StartSpeakerEnrollmentJob

See: newStartSpeakerEnrollmentJob smart constructor.

Constructors

StartSpeakerEnrollmentJob'

Fields

clientToken :: Maybe Text

The idempotency token for starting a new speaker enrollment Job. If not provided, Amazon Web Services SDK populates this field.

jobName :: Maybe (Sensitive Text)

A name for your speaker enrollment job.

enrollmentConfig :: Maybe EnrollmentConfig

The enrollment config that contains details such as the action to take when a speaker is already enrolled in the Voice ID system or when a speaker is identified as a fraudster.

dataAccessRoleArn :: Text

The IAM role Amazon Resource Name (ARN) that grants Voice ID permissions to access customer's buckets to read the input manifest file and write the job output file. Refer to Batch enrollment using audio data from prior calls documentation for the permissions needed in this role.

domainId :: Text

The identifier of the domain that contains the speaker enrollment job and in which the speakers are enrolled.

inputDataConfig :: InputDataConfig

The input data config containing the S3 location for the input manifest file that contains the list of speaker enrollment requests.

outputDataConfig :: OutputDataConfig

The output data config containing the S3 location where Voice ID writes the job output file; you must also include a KMS Key ID to encrypt the file.

newStartSpeakerEnrollmentJob

Arguments

:: Text
$sel:dataAccessRoleArn:StartSpeakerEnrollmentJob'

-> Text
$sel:domainId:StartSpeakerEnrollmentJob'

-> InputDataConfig
$sel:inputDataConfig:StartSpeakerEnrollmentJob'

-> OutputDataConfig
$sel:outputDataConfig:StartSpeakerEnrollmentJob'

-> StartSpeakerEnrollmentJob

Create a value of StartSpeakerEnrollmentJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:StartSpeakerEnrollmentJob', startSpeakerEnrollmentJob_clientToken - The idempotency token for starting a new speaker enrollment Job. If not provided, Amazon Web Services SDK populates this field.

$sel:jobName:StartSpeakerEnrollmentJob', startSpeakerEnrollmentJob_jobName - A name for your speaker enrollment job.

$sel:enrollmentConfig:StartSpeakerEnrollmentJob', startSpeakerEnrollmentJob_enrollmentConfig - The enrollment config that contains details such as the action to take when a speaker is already enrolled in the Voice ID system or when a speaker is identified as a fraudster.

$sel:dataAccessRoleArn:StartSpeakerEnrollmentJob', startSpeakerEnrollmentJob_dataAccessRoleArn - The IAM role Amazon Resource Name (ARN) that grants Voice ID permissions to access customer's buckets to read the input manifest file and write the job output file. Refer to Batch enrollment using audio data from prior calls documentation for the permissions needed in this role.

$sel:domainId:StartSpeakerEnrollmentJob', startSpeakerEnrollmentJob_domainId - The identifier of the domain that contains the speaker enrollment job and in which the speakers are enrolled.

$sel:inputDataConfig:StartSpeakerEnrollmentJob', startSpeakerEnrollmentJob_inputDataConfig - The input data config containing the S3 location for the input manifest file that contains the list of speaker enrollment requests.

$sel:outputDataConfig:StartSpeakerEnrollmentJob', startSpeakerEnrollmentJob_outputDataConfig - The output data config containing the S3 location where Voice ID writes the job output file; you must also include a KMS Key ID to encrypt the file.

Request Lenses

startSpeakerEnrollmentJob_clientToken :: Lens' StartSpeakerEnrollmentJob (Maybe Text)

The idempotency token for starting a new speaker enrollment Job. If not provided, Amazon Web Services SDK populates this field.

startSpeakerEnrollmentJob_jobName :: Lens' StartSpeakerEnrollmentJob (Maybe Text)

A name for your speaker enrollment job.

startSpeakerEnrollmentJob_enrollmentConfig :: Lens' StartSpeakerEnrollmentJob (Maybe EnrollmentConfig)

The enrollment config that contains details such as the action to take when a speaker is already enrolled in the Voice ID system or when a speaker is identified as a fraudster.

startSpeakerEnrollmentJob_dataAccessRoleArn :: Lens' StartSpeakerEnrollmentJob Text

The IAM role Amazon Resource Name (ARN) that grants Voice ID permissions to access customer's buckets to read the input manifest file and write the job output file. Refer to Batch enrollment using audio data from prior calls documentation for the permissions needed in this role.

startSpeakerEnrollmentJob_domainId :: Lens' StartSpeakerEnrollmentJob Text

The identifier of the domain that contains the speaker enrollment job and in which the speakers are enrolled.

startSpeakerEnrollmentJob_inputDataConfig :: Lens' StartSpeakerEnrollmentJob InputDataConfig

The input data config containing the S3 location for the input manifest file that contains the list of speaker enrollment requests.

startSpeakerEnrollmentJob_outputDataConfig :: Lens' StartSpeakerEnrollmentJob OutputDataConfig

The output data config containing the S3 location where Voice ID writes the job output file; you must also include a KMS Key ID to encrypt the file.

Destructuring the Response

data StartSpeakerEnrollmentJobResponse

See: newStartSpeakerEnrollmentJobResponse smart constructor.

Constructors

StartSpeakerEnrollmentJobResponse'

Fields

job :: Maybe SpeakerEnrollmentJob

Details about the started speaker enrollment job.

httpStatus :: Int

The response's http status code.

newStartSpeakerEnrollmentJobResponse

Arguments

:: Int
$sel:httpStatus:StartSpeakerEnrollmentJobResponse'

-> StartSpeakerEnrollmentJobResponse

Create a value of StartSpeakerEnrollmentJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:job:StartSpeakerEnrollmentJobResponse', startSpeakerEnrollmentJobResponse_job - Details about the started speaker enrollment job.

$sel:httpStatus:StartSpeakerEnrollmentJobResponse', startSpeakerEnrollmentJobResponse_httpStatus - The response's http status code.

Response Lenses

startSpeakerEnrollmentJobResponse_job :: Lens' StartSpeakerEnrollmentJobResponse (Maybe SpeakerEnrollmentJob)

Details about the started speaker enrollment job.

startSpeakerEnrollmentJobResponse_httpStatus :: Lens' StartSpeakerEnrollmentJobResponse Int

The response's http status code.