Network.AWS.VoiceId.Types.Domain

Description

Documentation

data Domain

Contains all the information about a domain.

See: newDomain smart constructor.

Constructors

Domain'

Fields

domainStatus :: Maybe DomainStatus

The current status of the domain.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the domain.

createdAt :: Maybe POSIX

The timestamp at which the domain is created.

name :: Maybe (Sensitive Text)

The client-provided name for the domain.

domainId :: Maybe Text

The service-generated identifier for the domain.

updatedAt :: Maybe POSIX

The timestamp showing the domain's last update.

description :: Maybe (Sensitive Text)

The client-provided description of the domain.

serverSideEncryptionConfiguration :: Maybe ServerSideEncryptionConfiguration

The server-side encryption configuration containing the KMS Key Identifier you want Voice ID to use to encrypt your data.

newDomain :: Domain

Create a value of Domain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainStatus:Domain', domain_domainStatus - The current status of the domain.

$sel:arn:Domain', domain_arn - The Amazon Resource Name (ARN) for the domain.

$sel:createdAt:Domain', domain_createdAt - The timestamp at which the domain is created.

$sel:name:Domain', domain_name - The client-provided name for the domain.

$sel:domainId:Domain', domain_domainId - The service-generated identifier for the domain.

$sel:updatedAt:Domain', domain_updatedAt - The timestamp showing the domain's last update.

$sel:description:Domain', domain_description - The client-provided description of the domain.

$sel:serverSideEncryptionConfiguration:Domain', domain_serverSideEncryptionConfiguration - The server-side encryption configuration containing the KMS Key Identifier you want Voice ID to use to encrypt your data.

domain_domainStatus :: Lens' Domain (Maybe DomainStatus)

The current status of the domain.

domain_arn :: Lens' Domain (Maybe Text)

The Amazon Resource Name (ARN) for the domain.

domain_createdAt :: Lens' Domain (Maybe UTCTime)

The timestamp at which the domain is created.

domain_name :: Lens' Domain (Maybe Text)

The client-provided name for the domain.

domain_domainId :: Lens' Domain (Maybe Text)

The service-generated identifier for the domain.

domain_updatedAt :: Lens' Domain (Maybe UTCTime)

The timestamp showing the domain's last update.

domain_description :: Lens' Domain (Maybe Text)

The client-provided description of the domain.

domain_serverSideEncryptionConfiguration :: Lens' Domain (Maybe ServerSideEncryptionConfiguration)

The server-side encryption configuration containing the KMS Key Identifier you want Voice ID to use to encrypt your data.