Network.AWS.VoiceId.Types.FraudsterRegistrationJobSummary

Description

Documentation

data FraudsterRegistrationJobSummary

Contains a summary of information about a fraudster registration job.

See: newFraudsterRegistrationJobSummary smart constructor.

Constructors

FraudsterRegistrationJobSummary'

Fields

failureDetails :: Maybe FailureDetails

Contains details that are populated when an entire batch job fails. In cases of individual registration job failures, the batch job as a whole doesn't fail; it is completed with a JobStatus of COMPLETED_WITH_ERRORS. You can use the job output file to identify the individual registration requests that failed.

jobId :: Maybe Text

The service-generated identifier for the fraudster registration job.

createdAt :: Maybe POSIX

A timestamp showing when the fraudster registration job is created.

jobName :: Maybe (Sensitive Text)

The client-provied name for the fraudster registration job.

endedAt :: Maybe POSIX

A timestamp showing when the fraudster registration job ended.

jobProgress :: Maybe JobProgress

Shows the completed percentage of registration requests listed in the input file.

domainId :: Maybe Text

The identifier of the domain containing the fraudster registration job.

jobStatus :: Maybe FraudsterRegistrationJobStatus

The current status of the fraudster registration job.

newFraudsterRegistrationJobSummary :: FraudsterRegistrationJobSummary

Create a value of FraudsterRegistrationJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureDetails:FraudsterRegistrationJobSummary', fraudsterRegistrationJobSummary_failureDetails - Contains details that are populated when an entire batch job fails. In cases of individual registration job failures, the batch job as a whole doesn't fail; it is completed with a JobStatus of COMPLETED_WITH_ERRORS. You can use the job output file to identify the individual registration requests that failed.

$sel:jobId:FraudsterRegistrationJobSummary', fraudsterRegistrationJobSummary_jobId - The service-generated identifier for the fraudster registration job.

$sel:createdAt:FraudsterRegistrationJobSummary', fraudsterRegistrationJobSummary_createdAt - A timestamp showing when the fraudster registration job is created.

$sel:jobName:FraudsterRegistrationJobSummary', fraudsterRegistrationJobSummary_jobName - The client-provied name for the fraudster registration job.

$sel:endedAt:FraudsterRegistrationJobSummary', fraudsterRegistrationJobSummary_endedAt - A timestamp showing when the fraudster registration job ended.

$sel:jobProgress:FraudsterRegistrationJobSummary', fraudsterRegistrationJobSummary_jobProgress - Shows the completed percentage of registration requests listed in the input file.

$sel:domainId:FraudsterRegistrationJobSummary', fraudsterRegistrationJobSummary_domainId - The identifier of the domain containing the fraudster registration job.

$sel:jobStatus:FraudsterRegistrationJobSummary', fraudsterRegistrationJobSummary_jobStatus - The current status of the fraudster registration job.

fraudsterRegistrationJobSummary_failureDetails :: Lens' FraudsterRegistrationJobSummary (Maybe FailureDetails)

Contains details that are populated when an entire batch job fails. In cases of individual registration job failures, the batch job as a whole doesn't fail; it is completed with a JobStatus of COMPLETED_WITH_ERRORS. You can use the job output file to identify the individual registration requests that failed.

fraudsterRegistrationJobSummary_jobId :: Lens' FraudsterRegistrationJobSummary (Maybe Text)

The service-generated identifier for the fraudster registration job.

fraudsterRegistrationJobSummary_createdAt :: Lens' FraudsterRegistrationJobSummary (Maybe UTCTime)

A timestamp showing when the fraudster registration job is created.

fraudsterRegistrationJobSummary_jobName :: Lens' FraudsterRegistrationJobSummary (Maybe Text)

The client-provied name for the fraudster registration job.

fraudsterRegistrationJobSummary_endedAt :: Lens' FraudsterRegistrationJobSummary (Maybe UTCTime)

A timestamp showing when the fraudster registration job ended.

fraudsterRegistrationJobSummary_jobProgress :: Lens' FraudsterRegistrationJobSummary (Maybe JobProgress)

Shows the completed percentage of registration requests listed in the input file.

fraudsterRegistrationJobSummary_domainId :: Lens' FraudsterRegistrationJobSummary (Maybe Text)

The identifier of the domain containing the fraudster registration job.

fraudsterRegistrationJobSummary_jobStatus :: Lens' FraudsterRegistrationJobSummary (Maybe FraudsterRegistrationJobStatus)

The current status of the fraudster registration job.