Network.AWS.WAF.CreateByteMatchSet

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Creates a ByteMatchSet. You then use UpdateByteMatchSet to identify the part of a web request that you want AWS WAF to inspect, such as the values of the User-Agent header or the query string. For example, you can create a ByteMatchSet that matches any requests with User-Agent headers that contain the string BadBot. You can then configure AWS WAF to reject those requests.

To create and configure a ByteMatchSet, perform the following steps:

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a CreateByteMatchSet request.

Submit a CreateByteMatchSet request.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateByteMatchSet request.

Submit an UpdateByteMatchSet request to specify the part of the request that you want AWS WAF to inspect (for example, the header or the URI) and the value that you want AWS WAF to watch for.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide.

Creating a Request

data CreateByteMatchSet

See: newCreateByteMatchSet smart constructor.

Constructors

CreateByteMatchSet'

Fields

name :: Text

A friendly name or description of the ByteMatchSet. You can't change Name after you create a ByteMatchSet.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

newCreateByteMatchSet

Arguments

:: Text
$sel:name:CreateByteMatchSet'

-> Text
$sel:changeToken:CreateByteMatchSet'

-> CreateByteMatchSet

Create a value of CreateByteMatchSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreateByteMatchSet', createByteMatchSet_name - A friendly name or description of the ByteMatchSet. You can't change Name after you create a ByteMatchSet.

$sel:changeToken:CreateByteMatchSet', createByteMatchSet_changeToken - The value returned by the most recent call to GetChangeToken.

Request Lenses

createByteMatchSet_name :: Lens' CreateByteMatchSet Text

A friendly name or description of the ByteMatchSet. You can't change Name after you create a ByteMatchSet.

createByteMatchSet_changeToken :: Lens' CreateByteMatchSet Text

The value returned by the most recent call to GetChangeToken.

Destructuring the Response

data CreateByteMatchSetResponse

See: newCreateByteMatchSetResponse smart constructor.

Constructors

CreateByteMatchSetResponse'

Fields

byteMatchSet :: Maybe ByteMatchSet

A ByteMatchSet that contains no ByteMatchTuple objects.

changeToken :: Maybe Text

The ChangeToken that you used to submit the CreateByteMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newCreateByteMatchSetResponse

Arguments

:: Int
$sel:httpStatus:CreateByteMatchSetResponse'

-> CreateByteMatchSetResponse

Create a value of CreateByteMatchSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:byteMatchSet:CreateByteMatchSetResponse', createByteMatchSetResponse_byteMatchSet - A ByteMatchSet that contains no ByteMatchTuple objects.

$sel:changeToken:CreateByteMatchSet', createByteMatchSetResponse_changeToken - The ChangeToken that you used to submit the CreateByteMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:CreateByteMatchSetResponse', createByteMatchSetResponse_httpStatus - The response's http status code.

Response Lenses

createByteMatchSetResponse_byteMatchSet :: Lens' CreateByteMatchSetResponse (Maybe ByteMatchSet)

A ByteMatchSet that contains no ByteMatchTuple objects.

createByteMatchSetResponse_changeToken :: Lens' CreateByteMatchSetResponse (Maybe Text)

The ChangeToken that you used to submit the CreateByteMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

createByteMatchSetResponse_httpStatus :: Lens' CreateByteMatchSetResponse Int

The response's http status code.