Network.AWS.WAF.CreateGeoMatchSet

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Creates an GeoMatchSet, which you use to specify which web requests you want to allow or block based on the country that the requests originate from. For example, if you're receiving a lot of requests from one or more countries and you want to block the requests, you can create an GeoMatchSet that contains those countries and then configure AWS WAF to block the requests.

To create and configure a GeoMatchSet, perform the following steps:

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a CreateGeoMatchSet request.

Submit a CreateGeoMatchSet request.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateGeoMatchSet request.

Submit an UpdateGeoMatchSetSet request to specify the countries that you want AWS WAF to watch for.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide.

Creating a Request

data CreateGeoMatchSet

See: newCreateGeoMatchSet smart constructor.

Constructors

CreateGeoMatchSet'

Fields

name :: Text

A friendly name or description of the GeoMatchSet. You can't change Name after you create the GeoMatchSet.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

newCreateGeoMatchSet

Arguments

:: Text
$sel:name:CreateGeoMatchSet'

-> Text
$sel:changeToken:CreateGeoMatchSet'

-> CreateGeoMatchSet

Create a value of CreateGeoMatchSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreateGeoMatchSet', createGeoMatchSet_name - A friendly name or description of the GeoMatchSet. You can't change Name after you create the GeoMatchSet.

$sel:changeToken:CreateGeoMatchSet', createGeoMatchSet_changeToken - The value returned by the most recent call to GetChangeToken.

Request Lenses

createGeoMatchSet_name :: Lens' CreateGeoMatchSet Text

A friendly name or description of the GeoMatchSet. You can't change Name after you create the GeoMatchSet.

createGeoMatchSet_changeToken :: Lens' CreateGeoMatchSet Text

The value returned by the most recent call to GetChangeToken.

Destructuring the Response

data CreateGeoMatchSetResponse

See: newCreateGeoMatchSetResponse smart constructor.

Constructors

CreateGeoMatchSetResponse'

Fields

geoMatchSet :: Maybe GeoMatchSet

The GeoMatchSet returned in the CreateGeoMatchSet response. The GeoMatchSet contains no GeoMatchConstraints.

changeToken :: Maybe Text

The ChangeToken that you used to submit the CreateGeoMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newCreateGeoMatchSetResponse

Arguments

:: Int
$sel:httpStatus:CreateGeoMatchSetResponse'

-> CreateGeoMatchSetResponse

Create a value of CreateGeoMatchSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:geoMatchSet:CreateGeoMatchSetResponse', createGeoMatchSetResponse_geoMatchSet - The GeoMatchSet returned in the CreateGeoMatchSet response. The GeoMatchSet contains no GeoMatchConstraints.

$sel:changeToken:CreateGeoMatchSet', createGeoMatchSetResponse_changeToken - The ChangeToken that you used to submit the CreateGeoMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:CreateGeoMatchSetResponse', createGeoMatchSetResponse_httpStatus - The response's http status code.

Response Lenses

createGeoMatchSetResponse_geoMatchSet :: Lens' CreateGeoMatchSetResponse (Maybe GeoMatchSet)

The GeoMatchSet returned in the CreateGeoMatchSet response. The GeoMatchSet contains no GeoMatchConstraints.

createGeoMatchSetResponse_changeToken :: Lens' CreateGeoMatchSetResponse (Maybe Text)

The ChangeToken that you used to submit the CreateGeoMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

createGeoMatchSetResponse_httpStatus :: Lens' CreateGeoMatchSetResponse Int

The response's http status code.