Network.AWS.WAF.CreateIPSet

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Creates an IPSet, which you use to specify which web requests that you want to allow or block based on the IP addresses that the requests originate from. For example, if you're receiving a lot of requests from one or more individual IP addresses or one or more ranges of IP addresses and you want to block the requests, you can create an IPSet that contains those IP addresses and then configure AWS WAF to block the requests.

To create and configure an IPSet, perform the following steps:

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a CreateIPSet request.

Submit a CreateIPSet request.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateIPSet request.

Submit an UpdateIPSet request to specify the IP addresses that you want AWS WAF to watch for.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide.

Creating a Request

data CreateIPSet

See: newCreateIPSet smart constructor.

Constructors

CreateIPSet'

Fields

name :: Text

A friendly name or description of the IPSet. You can't change Name after you create the IPSet.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

newCreateIPSet

Arguments

:: Text
$sel:name:CreateIPSet'

-> Text
$sel:changeToken:CreateIPSet'

-> CreateIPSet

Create a value of CreateIPSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreateIPSet', createIPSet_name - A friendly name or description of the IPSet. You can't change Name after you create the IPSet.

$sel:changeToken:CreateIPSet', createIPSet_changeToken - The value returned by the most recent call to GetChangeToken.

Request Lenses

createIPSet_name :: Lens' CreateIPSet Text

A friendly name or description of the IPSet. You can't change Name after you create the IPSet.

createIPSet_changeToken :: Lens' CreateIPSet Text

The value returned by the most recent call to GetChangeToken.

Destructuring the Response

data CreateIPSetResponse

See: newCreateIPSetResponse smart constructor.

Constructors

CreateIPSetResponse'

Fields

changeToken :: Maybe Text

The ChangeToken that you used to submit the CreateIPSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

iPSet :: Maybe IPSet

The IPSet returned in the CreateIPSet response.

httpStatus :: Int

The response's http status code.

newCreateIPSetResponse

Arguments

:: Int
$sel:httpStatus:CreateIPSetResponse'

-> CreateIPSetResponse

Create a value of CreateIPSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeToken:CreateIPSet', createIPSetResponse_changeToken - The ChangeToken that you used to submit the CreateIPSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:iPSet:CreateIPSetResponse', createIPSetResponse_iPSet - The IPSet returned in the CreateIPSet response.

$sel:httpStatus:CreateIPSetResponse', createIPSetResponse_httpStatus - The response's http status code.

Response Lenses

createIPSetResponse_changeToken :: Lens' CreateIPSetResponse (Maybe Text)

The ChangeToken that you used to submit the CreateIPSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

createIPSetResponse_iPSet :: Lens' CreateIPSetResponse (Maybe IPSet)

The IPSet returned in the CreateIPSet response.

createIPSetResponse_httpStatus :: Lens' CreateIPSetResponse Int

The response's http status code.