Network.AWS.WAF.CreateRateBasedRule

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Creates a RateBasedRule. The RateBasedRule contains a RateLimit, which specifies the maximum number of requests that AWS WAF allows from a specified IP address in a five-minute period. The RateBasedRule also contains the IPSet objects, ByteMatchSet objects, and other predicates that identify the requests that you want to count or block if these requests exceed the RateLimit.

If you add more than one predicate to a RateBasedRule, a request not only must exceed the RateLimit, but it also must match all the conditions to be counted or blocked. For example, suppose you add the following to a RateBasedRule:

An IPSet that matches the IP address 192.0.2.44/32

A ByteMatchSet that matches BadBot in the User-Agent header

Further, you specify a RateLimit of 1,000.

You then add the RateBasedRule to a WebACL and specify that you want to block requests that meet the conditions in the rule. For a request to be blocked, it must come from the IP address 192.0.2.44 and the User-Agent header in the request must contain the value BadBot. Further, requests that match these two conditions must be received at a rate of more than 1,000 requests every five minutes. If both conditions are met and the rate is exceeded, AWS WAF blocks the requests. If the rate drops below 1,000 for a five-minute period, AWS WAF no longer blocks the requests.

As a second example, suppose you want to limit requests to a particular page on your site. To do this, you could add the following to a RateBasedRule:

A ByteMatchSet with FieldToMatch of URI

A PositionalConstraint of STARTS_WITH

A TargetString of login

Further, you specify a RateLimit of 1,000.

By adding this RateBasedRule to a WebACL, you could limit requests to your login page without affecting the rest of your site.

To create and configure a RateBasedRule, perform the following steps:

Create and update the predicates that you want to include in the rule. For more information, see CreateByteMatchSet, CreateIPSet, and CreateSqlInjectionMatchSet.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a CreateRule request.

Submit a CreateRateBasedRule request.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateRule request.

Submit an UpdateRateBasedRule request to specify the predicates that you want to include in the rule.

Create and update a WebACL that contains the RateBasedRule. For more information, see CreateWebACL.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide.

Creating a Request

data CreateRateBasedRule

See: newCreateRateBasedRule smart constructor.

Constructors

CreateRateBasedRule'

Fields

tags :: Maybe (NonEmpty Tag)

name :: Text

A friendly name or description of the RateBasedRule. You can't change the name of a RateBasedRule after you create it.

metricName :: Text

A friendly name or description for the metrics for this RateBasedRule. The name can contain only alphanumeric characters (A-Z, a-z, 0-9), with maximum length 128 and minimum length one. It can't contain whitespace or metric names reserved for AWS WAF, including "All" and "Default_Action." You can't change the name of the metric after you create the RateBasedRule.

rateKey :: RateKey

The field that AWS WAF uses to determine if requests are likely arriving from a single source and thus subject to rate monitoring. The only valid value for RateKey is IP. IP indicates that requests that arrive from the same IP address are subject to the RateLimit that is specified in the RateBasedRule.

rateLimit :: Natural

The maximum number of requests, which have an identical value in the field that is specified by RateKey, allowed in a five-minute period. If the number of requests exceeds the RateLimit and the other predicates specified in the rule are also met, AWS WAF triggers the action that is specified for this rule.

changeToken :: Text

The ChangeToken that you used to submit the CreateRateBasedRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

newCreateRateBasedRule

Arguments

:: Text
$sel:name:CreateRateBasedRule'

-> Text
$sel:metricName:CreateRateBasedRule'

-> RateKey
$sel:rateKey:CreateRateBasedRule'

-> Natural
$sel:rateLimit:CreateRateBasedRule'

-> Text
$sel:changeToken:CreateRateBasedRule'

-> CreateRateBasedRule

Create a value of CreateRateBasedRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateRateBasedRule', createRateBasedRule_tags -

$sel:name:CreateRateBasedRule', createRateBasedRule_name - A friendly name or description of the RateBasedRule. You can't change the name of a RateBasedRule after you create it.

$sel:metricName:CreateRateBasedRule', createRateBasedRule_metricName - A friendly name or description for the metrics for this RateBasedRule. The name can contain only alphanumeric characters (A-Z, a-z, 0-9), with maximum length 128 and minimum length one. It can't contain whitespace or metric names reserved for AWS WAF, including "All" and "Default_Action." You can't change the name of the metric after you create the RateBasedRule.

$sel:rateKey:CreateRateBasedRule', createRateBasedRule_rateKey - The field that AWS WAF uses to determine if requests are likely arriving from a single source and thus subject to rate monitoring. The only valid value for RateKey is IP. IP indicates that requests that arrive from the same IP address are subject to the RateLimit that is specified in the RateBasedRule.

$sel:rateLimit:CreateRateBasedRule', createRateBasedRule_rateLimit - The maximum number of requests, which have an identical value in the field that is specified by RateKey, allowed in a five-minute period. If the number of requests exceeds the RateLimit and the other predicates specified in the rule are also met, AWS WAF triggers the action that is specified for this rule.

$sel:changeToken:CreateRateBasedRule', createRateBasedRule_changeToken - The ChangeToken that you used to submit the CreateRateBasedRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

Request Lenses

createRateBasedRule_tags :: Lens' CreateRateBasedRule (Maybe (NonEmpty Tag))

createRateBasedRule_name :: Lens' CreateRateBasedRule Text

A friendly name or description of the RateBasedRule. You can't change the name of a RateBasedRule after you create it.

createRateBasedRule_metricName :: Lens' CreateRateBasedRule Text

A friendly name or description for the metrics for this RateBasedRule. The name can contain only alphanumeric characters (A-Z, a-z, 0-9), with maximum length 128 and minimum length one. It can't contain whitespace or metric names reserved for AWS WAF, including "All" and "Default_Action." You can't change the name of the metric after you create the RateBasedRule.

createRateBasedRule_rateKey :: Lens' CreateRateBasedRule RateKey

The field that AWS WAF uses to determine if requests are likely arriving from a single source and thus subject to rate monitoring. The only valid value for RateKey is IP. IP indicates that requests that arrive from the same IP address are subject to the RateLimit that is specified in the RateBasedRule.

createRateBasedRule_rateLimit :: Lens' CreateRateBasedRule Natural

The maximum number of requests, which have an identical value in the field that is specified by RateKey, allowed in a five-minute period. If the number of requests exceeds the RateLimit and the other predicates specified in the rule are also met, AWS WAF triggers the action that is specified for this rule.

createRateBasedRule_changeToken :: Lens' CreateRateBasedRule Text

The ChangeToken that you used to submit the CreateRateBasedRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

Destructuring the Response

data CreateRateBasedRuleResponse

See: newCreateRateBasedRuleResponse smart constructor.

Constructors

CreateRateBasedRuleResponse'

Fields

rule :: Maybe RateBasedRule

The RateBasedRule that is returned in the CreateRateBasedRule response.

changeToken :: Maybe Text

The ChangeToken that you used to submit the CreateRateBasedRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newCreateRateBasedRuleResponse

Arguments

:: Int
$sel:httpStatus:CreateRateBasedRuleResponse'

-> CreateRateBasedRuleResponse

Create a value of CreateRateBasedRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rule:CreateRateBasedRuleResponse', createRateBasedRuleResponse_rule - The RateBasedRule that is returned in the CreateRateBasedRule response.

$sel:changeToken:CreateRateBasedRule', createRateBasedRuleResponse_changeToken - The ChangeToken that you used to submit the CreateRateBasedRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:CreateRateBasedRuleResponse', createRateBasedRuleResponse_httpStatus - The response's http status code.

Response Lenses

createRateBasedRuleResponse_rule :: Lens' CreateRateBasedRuleResponse (Maybe RateBasedRule)

The RateBasedRule that is returned in the CreateRateBasedRule response.

createRateBasedRuleResponse_changeToken :: Lens' CreateRateBasedRuleResponse (Maybe Text)

The ChangeToken that you used to submit the CreateRateBasedRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

createRateBasedRuleResponse_httpStatus :: Lens' CreateRateBasedRuleResponse Int

The response's http status code.