Network.AWS.WAF.DeleteRegexMatchSet

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Permanently deletes a RegexMatchSet. You can't delete a RegexMatchSet if it's still used in any Rules or if it still includes any RegexMatchTuples objects (any filters).

If you just want to remove a RegexMatchSet from a Rule, use UpdateRule.

To permanently delete a RegexMatchSet, perform the following steps:

Update the RegexMatchSet to remove filters, if any. For more information, see UpdateRegexMatchSet.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a DeleteRegexMatchSet request.

Submit a DeleteRegexMatchSet request.

Creating a Request

data DeleteRegexMatchSet

See: newDeleteRegexMatchSet smart constructor.

Constructors

DeleteRegexMatchSet'

Fields

regexMatchSetId :: Text

The RegexMatchSetId of the RegexMatchSet that you want to delete. RegexMatchSetId is returned by CreateRegexMatchSet and by ListRegexMatchSets.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

newDeleteRegexMatchSet

Arguments

:: Text
$sel:regexMatchSetId:DeleteRegexMatchSet'

-> Text
$sel:changeToken:DeleteRegexMatchSet'

-> DeleteRegexMatchSet

Create a value of DeleteRegexMatchSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:regexMatchSetId:DeleteRegexMatchSet', deleteRegexMatchSet_regexMatchSetId - The RegexMatchSetId of the RegexMatchSet that you want to delete. RegexMatchSetId is returned by CreateRegexMatchSet and by ListRegexMatchSets.

$sel:changeToken:DeleteRegexMatchSet', deleteRegexMatchSet_changeToken - The value returned by the most recent call to GetChangeToken.

Request Lenses

deleteRegexMatchSet_regexMatchSetId :: Lens' DeleteRegexMatchSet Text

The RegexMatchSetId of the RegexMatchSet that you want to delete. RegexMatchSetId is returned by CreateRegexMatchSet and by ListRegexMatchSets.

deleteRegexMatchSet_changeToken :: Lens' DeleteRegexMatchSet Text

The value returned by the most recent call to GetChangeToken.

Destructuring the Response

data DeleteRegexMatchSetResponse

See: newDeleteRegexMatchSetResponse smart constructor.

Constructors

DeleteRegexMatchSetResponse'

Fields

changeToken :: Maybe Text

The ChangeToken that you used to submit the DeleteRegexMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newDeleteRegexMatchSetResponse

Arguments

:: Int
$sel:httpStatus:DeleteRegexMatchSetResponse'

-> DeleteRegexMatchSetResponse

Create a value of DeleteRegexMatchSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeToken:DeleteRegexMatchSet', deleteRegexMatchSetResponse_changeToken - The ChangeToken that you used to submit the DeleteRegexMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:DeleteRegexMatchSetResponse', deleteRegexMatchSetResponse_httpStatus - The response's http status code.

Response Lenses

deleteRegexMatchSetResponse_changeToken :: Lens' DeleteRegexMatchSetResponse (Maybe Text)

The ChangeToken that you used to submit the DeleteRegexMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

deleteRegexMatchSetResponse_httpStatus :: Lens' DeleteRegexMatchSetResponse Int

The response's http status code.