Network.AWS.WAF.DeleteRule

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Permanently deletes a Rule. You can't delete a Rule if it's still used in any WebACL objects or if it still includes any predicates, such as ByteMatchSet objects.

If you just want to remove a Rule from a WebACL, use UpdateWebACL.

To permanently delete a Rule from AWS WAF, perform the following steps:

Update the Rule to remove predicates, if any. For more information, see UpdateRule.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a DeleteRule request.

Submit a DeleteRule request.

Creating a Request

data DeleteRule

See: newDeleteRule smart constructor.

Constructors

DeleteRule'

Fields

ruleId :: Text

The RuleId of the Rule that you want to delete. RuleId is returned by CreateRule and by ListRules.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

newDeleteRule

Arguments

:: Text
$sel:ruleId:DeleteRule'

-> Text
$sel:changeToken:DeleteRule'

-> DeleteRule

Create a value of DeleteRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleId:DeleteRule', deleteRule_ruleId - The RuleId of the Rule that you want to delete. RuleId is returned by CreateRule and by ListRules.

$sel:changeToken:DeleteRule', deleteRule_changeToken - The value returned by the most recent call to GetChangeToken.

Request Lenses

deleteRule_ruleId :: Lens' DeleteRule Text

The RuleId of the Rule that you want to delete. RuleId is returned by CreateRule and by ListRules.

deleteRule_changeToken :: Lens' DeleteRule Text

The value returned by the most recent call to GetChangeToken.

Destructuring the Response

data DeleteRuleResponse

See: newDeleteRuleResponse smart constructor.

Constructors

DeleteRuleResponse'

Fields

changeToken :: Maybe Text

The ChangeToken that you used to submit the DeleteRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newDeleteRuleResponse

Arguments

:: Int
$sel:httpStatus:DeleteRuleResponse'

-> DeleteRuleResponse

Create a value of DeleteRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeToken:DeleteRule', deleteRuleResponse_changeToken - The ChangeToken that you used to submit the DeleteRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:DeleteRuleResponse', deleteRuleResponse_httpStatus - The response's http status code.

Response Lenses

deleteRuleResponse_changeToken :: Lens' DeleteRuleResponse (Maybe Text)

The ChangeToken that you used to submit the DeleteRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

deleteRuleResponse_httpStatus :: Lens' DeleteRuleResponse Int

The response's http status code.