Network.AWS.WAF.GetRateBasedRuleManagedKeys

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Returns an array of IP addresses currently being blocked by the RateBasedRule that is specified by the RuleId. The maximum number of managed keys that will be blocked is 10,000. If more than 10,000 addresses exceed the rate limit, the 10,000 addresses with the highest rates will be blocked.

This operation returns paginated results.

Creating a Request

data GetRateBasedRuleManagedKeys

See: newGetRateBasedRuleManagedKeys smart constructor.

Constructors

GetRateBasedRuleManagedKeys'

Fields

nextMarker :: Maybe Text

A null value and not currently used. Do not include this in your request.

ruleId :: Text

The RuleId of the RateBasedRule for which you want to get a list of ManagedKeys. RuleId is returned by CreateRateBasedRule and by ListRateBasedRules.

newGetRateBasedRuleManagedKeys

Arguments

:: Text
$sel:ruleId:GetRateBasedRuleManagedKeys'

-> GetRateBasedRuleManagedKeys

Create a value of GetRateBasedRuleManagedKeys with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextMarker:GetRateBasedRuleManagedKeys', getRateBasedRuleManagedKeys_nextMarker - A null value and not currently used. Do not include this in your request.

$sel:ruleId:GetRateBasedRuleManagedKeys', getRateBasedRuleManagedKeys_ruleId - The RuleId of the RateBasedRule for which you want to get a list of ManagedKeys. RuleId is returned by CreateRateBasedRule and by ListRateBasedRules.

Request Lenses

getRateBasedRuleManagedKeys_nextMarker :: Lens' GetRateBasedRuleManagedKeys (Maybe Text)

A null value and not currently used. Do not include this in your request.

getRateBasedRuleManagedKeys_ruleId :: Lens' GetRateBasedRuleManagedKeys Text

The RuleId of the RateBasedRule for which you want to get a list of ManagedKeys. RuleId is returned by CreateRateBasedRule and by ListRateBasedRules.

Destructuring the Response

data GetRateBasedRuleManagedKeysResponse

See: newGetRateBasedRuleManagedKeysResponse smart constructor.

Constructors

GetRateBasedRuleManagedKeysResponse'

Fields

nextMarker :: Maybe Text

A null value and not currently used.

managedKeys :: Maybe [Text]

An array of IP addresses that currently are blocked by the specified RateBasedRule.

httpStatus :: Int

The response's http status code.

newGetRateBasedRuleManagedKeysResponse

Arguments

:: Int
$sel:httpStatus:GetRateBasedRuleManagedKeysResponse'

-> GetRateBasedRuleManagedKeysResponse

Create a value of GetRateBasedRuleManagedKeysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextMarker:GetRateBasedRuleManagedKeys', getRateBasedRuleManagedKeysResponse_nextMarker - A null value and not currently used.

$sel:managedKeys:GetRateBasedRuleManagedKeysResponse', getRateBasedRuleManagedKeysResponse_managedKeys - An array of IP addresses that currently are blocked by the specified RateBasedRule.

$sel:httpStatus:GetRateBasedRuleManagedKeysResponse', getRateBasedRuleManagedKeysResponse_httpStatus - The response's http status code.

Response Lenses

getRateBasedRuleManagedKeysResponse_nextMarker :: Lens' GetRateBasedRuleManagedKeysResponse (Maybe Text)

A null value and not currently used.

getRateBasedRuleManagedKeysResponse_managedKeys :: Lens' GetRateBasedRuleManagedKeysResponse (Maybe [Text])

An array of IP addresses that currently are blocked by the specified RateBasedRule.

getRateBasedRuleManagedKeysResponse_httpStatus :: Lens' GetRateBasedRuleManagedKeysResponse Int

The response's http status code.