Network.AWS.WAF.GetSampledRequests

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Gets detailed information about a specified number of requests--a sample--that AWS WAF randomly selects from among the first 5,000 requests that your AWS resource received during a time range that you choose. You can specify a sample size of up to 500 requests, and you can specify any time range in the previous three hours.

GetSampledRequests returns a time range, which is usually the time range that you specified. However, if your resource (such as a CloudFront distribution) received 5,000 requests before the specified time range elapsed, GetSampledRequests returns an updated time range. This new time range indicates the actual period during which AWS WAF selected the requests in the sample.

Creating a Request

data GetSampledRequests

See: newGetSampledRequests smart constructor.

Constructors

GetSampledRequests'

Fields

webAclId :: Text

The WebACLId of the WebACL for which you want GetSampledRequests to return a sample of requests.

ruleId :: Text

RuleId is one of three values:

The RuleId of the Rule or the RuleGroupId of the RuleGroup for which you want GetSampledRequests to return a sample of requests.

Default_Action, which causes GetSampledRequests to return a sample of the requests that didn't match any of the rules in the specified WebACL.

timeWindow :: TimeWindow

The start date and time and the end date and time of the range for which you want GetSampledRequests to return a sample of requests. You must specify the times in Coordinated Universal Time (UTC) format. UTC format includes the special designator, Z. For example, "2016-09-27T14:50Z". You can specify any time range in the previous three hours.

maxItems :: Natural

The number of requests that you want AWS WAF to return from among the first 5,000 requests that your AWS resource received during the time range. If your resource received fewer requests than the value of MaxItems, GetSampledRequests returns information about all of them.

newGetSampledRequests

Arguments

:: Text
$sel:webAclId:GetSampledRequests'

-> Text
$sel:ruleId:GetSampledRequests'

-> TimeWindow
$sel:timeWindow:GetSampledRequests'

-> Natural
$sel:maxItems:GetSampledRequests'

-> GetSampledRequests

Create a value of GetSampledRequests with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:webAclId:GetSampledRequests', getSampledRequests_webAclId - The WebACLId of the WebACL for which you want GetSampledRequests to return a sample of requests.

$sel:ruleId:GetSampledRequests', getSampledRequests_ruleId - RuleId is one of three values:

The RuleId of the Rule or the RuleGroupId of the RuleGroup for which you want GetSampledRequests to return a sample of requests.

Default_Action, which causes GetSampledRequests to return a sample of the requests that didn't match any of the rules in the specified WebACL.

$sel:timeWindow:GetSampledRequests', getSampledRequests_timeWindow - The start date and time and the end date and time of the range for which you want GetSampledRequests to return a sample of requests. You must specify the times in Coordinated Universal Time (UTC) format. UTC format includes the special designator, Z. For example, "2016-09-27T14:50Z". You can specify any time range in the previous three hours.

$sel:maxItems:GetSampledRequests', getSampledRequests_maxItems - The number of requests that you want AWS WAF to return from among the first 5,000 requests that your AWS resource received during the time range. If your resource received fewer requests than the value of MaxItems, GetSampledRequests returns information about all of them.

Request Lenses

getSampledRequests_webAclId :: Lens' GetSampledRequests Text

The WebACLId of the WebACL for which you want GetSampledRequests to return a sample of requests.

getSampledRequests_ruleId :: Lens' GetSampledRequests Text

RuleId is one of three values:

The RuleId of the Rule or the RuleGroupId of the RuleGroup for which you want GetSampledRequests to return a sample of requests.

Default_Action, which causes GetSampledRequests to return a sample of the requests that didn't match any of the rules in the specified WebACL.

getSampledRequests_timeWindow :: Lens' GetSampledRequests TimeWindow

The start date and time and the end date and time of the range for which you want GetSampledRequests to return a sample of requests. You must specify the times in Coordinated Universal Time (UTC) format. UTC format includes the special designator, Z. For example, "2016-09-27T14:50Z". You can specify any time range in the previous three hours.

getSampledRequests_maxItems :: Lens' GetSampledRequests Natural

The number of requests that you want AWS WAF to return from among the first 5,000 requests that your AWS resource received during the time range. If your resource received fewer requests than the value of MaxItems, GetSampledRequests returns information about all of them.

Destructuring the Response

data GetSampledRequestsResponse

See: newGetSampledRequestsResponse smart constructor.

Constructors

GetSampledRequestsResponse'

Fields

timeWindow :: Maybe TimeWindow

Usually, TimeWindow is the time range that you specified in the GetSampledRequests request. However, if your AWS resource received more than 5,000 requests during the time range that you specified in the request, GetSampledRequests returns the time range for the first 5,000 requests. Times are in Coordinated Universal Time (UTC) format.

populationSize :: Maybe Integer

The total number of requests from which GetSampledRequests got a sample of MaxItems requests. If PopulationSize is less than MaxItems, the sample includes every request that your AWS resource received during the specified time range.

sampledRequests :: Maybe [SampledHTTPRequest]

A complex type that contains detailed information about each of the requests in the sample.

httpStatus :: Int

The response's http status code.

newGetSampledRequestsResponse

Arguments

:: Int
$sel:httpStatus:GetSampledRequestsResponse'

-> GetSampledRequestsResponse

Create a value of GetSampledRequestsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeWindow:GetSampledRequests', getSampledRequestsResponse_timeWindow - Usually, TimeWindow is the time range that you specified in the GetSampledRequests request. However, if your AWS resource received more than 5,000 requests during the time range that you specified in the request, GetSampledRequests returns the time range for the first 5,000 requests. Times are in Coordinated Universal Time (UTC) format.

$sel:populationSize:GetSampledRequestsResponse', getSampledRequestsResponse_populationSize - The total number of requests from which GetSampledRequests got a sample of MaxItems requests. If PopulationSize is less than MaxItems, the sample includes every request that your AWS resource received during the specified time range.

$sel:sampledRequests:GetSampledRequestsResponse', getSampledRequestsResponse_sampledRequests - A complex type that contains detailed information about each of the requests in the sample.

$sel:httpStatus:GetSampledRequestsResponse', getSampledRequestsResponse_httpStatus - The response's http status code.

Response Lenses

getSampledRequestsResponse_timeWindow :: Lens' GetSampledRequestsResponse (Maybe TimeWindow)

Usually, TimeWindow is the time range that you specified in the GetSampledRequests request. However, if your AWS resource received more than 5,000 requests during the time range that you specified in the request, GetSampledRequests returns the time range for the first 5,000 requests. Times are in Coordinated Universal Time (UTC) format.

getSampledRequestsResponse_populationSize :: Lens' GetSampledRequestsResponse (Maybe Integer)

The total number of requests from which GetSampledRequests got a sample of MaxItems requests. If PopulationSize is less than MaxItems, the sample includes every request that your AWS resource received during the specified time range.

getSampledRequestsResponse_sampledRequests :: Lens' GetSampledRequestsResponse (Maybe [SampledHTTPRequest])

A complex type that contains detailed information about each of the requests in the sample.

getSampledRequestsResponse_httpStatus :: Lens' GetSampledRequestsResponse Int

The response's http status code.