Network.AWS.WAF.ListByteMatchSets

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Returns an array of ByteMatchSetSummary objects.

This operation returns paginated results.

Creating a Request

data ListByteMatchSets

See: newListByteMatchSets smart constructor.

Constructors

ListByteMatchSets'

Fields

nextMarker :: Maybe Text

If you specify a value for Limit and you have more ByteMatchSets than the value of Limit, AWS WAF returns a NextMarker value in the response that allows you to list another group of ByteMatchSets. For the second and subsequent ListByteMatchSets requests, specify the value of NextMarker from the previous response to get information about another batch of ByteMatchSets.

limit :: Maybe Natural

Specifies the number of ByteMatchSet objects that you want AWS WAF to return for this request. If you have more ByteMatchSets objects than the number you specify for Limit, the response includes a NextMarker value that you can use to get another batch of ByteMatchSet objects.

newListByteMatchSets :: ListByteMatchSets

Create a value of ListByteMatchSets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextMarker:ListByteMatchSets', listByteMatchSets_nextMarker - If you specify a value for Limit and you have more ByteMatchSets than the value of Limit, AWS WAF returns a NextMarker value in the response that allows you to list another group of ByteMatchSets. For the second and subsequent ListByteMatchSets requests, specify the value of NextMarker from the previous response to get information about another batch of ByteMatchSets.

$sel:limit:ListByteMatchSets', listByteMatchSets_limit - Specifies the number of ByteMatchSet objects that you want AWS WAF to return for this request. If you have more ByteMatchSets objects than the number you specify for Limit, the response includes a NextMarker value that you can use to get another batch of ByteMatchSet objects.

Request Lenses

listByteMatchSets_nextMarker :: Lens' ListByteMatchSets (Maybe Text)

If you specify a value for Limit and you have more ByteMatchSets than the value of Limit, AWS WAF returns a NextMarker value in the response that allows you to list another group of ByteMatchSets. For the second and subsequent ListByteMatchSets requests, specify the value of NextMarker from the previous response to get information about another batch of ByteMatchSets.

listByteMatchSets_limit :: Lens' ListByteMatchSets (Maybe Natural)

Specifies the number of ByteMatchSet objects that you want AWS WAF to return for this request. If you have more ByteMatchSets objects than the number you specify for Limit, the response includes a NextMarker value that you can use to get another batch of ByteMatchSet objects.

Destructuring the Response

data ListByteMatchSetsResponse

See: newListByteMatchSetsResponse smart constructor.

Constructors

ListByteMatchSetsResponse'

Fields

nextMarker :: Maybe Text

If you have more ByteMatchSet objects than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more ByteMatchSet objects, submit another ListByteMatchSets request, and specify the NextMarker value from the response in the NextMarker value in the next request.

byteMatchSets :: Maybe [ByteMatchSetSummary]

An array of ByteMatchSetSummary objects.

httpStatus :: Int

The response's http status code.

newListByteMatchSetsResponse

Arguments

:: Int
$sel:httpStatus:ListByteMatchSetsResponse'

-> ListByteMatchSetsResponse

Create a value of ListByteMatchSetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextMarker:ListByteMatchSets', listByteMatchSetsResponse_nextMarker - If you have more ByteMatchSet objects than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more ByteMatchSet objects, submit another ListByteMatchSets request, and specify the NextMarker value from the response in the NextMarker value in the next request.

$sel:byteMatchSets:ListByteMatchSetsResponse', listByteMatchSetsResponse_byteMatchSets - An array of ByteMatchSetSummary objects.

$sel:httpStatus:ListByteMatchSetsResponse', listByteMatchSetsResponse_httpStatus - The response's http status code.

Response Lenses

listByteMatchSetsResponse_nextMarker :: Lens' ListByteMatchSetsResponse (Maybe Text)

If you have more ByteMatchSet objects than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more ByteMatchSet objects, submit another ListByteMatchSets request, and specify the NextMarker value from the response in the NextMarker value in the next request.

listByteMatchSetsResponse_byteMatchSets :: Lens' ListByteMatchSetsResponse (Maybe [ByteMatchSetSummary])

An array of ByteMatchSetSummary objects.

listByteMatchSetsResponse_httpStatus :: Lens' ListByteMatchSetsResponse Int

The response's http status code.