Network.AWS.WAF.PutPermissionPolicy

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Attaches an IAM policy to the specified resource. The only supported use for this action is to share a RuleGroup across accounts.

The PutPermissionPolicy is subject to the following restrictions:

You can attach only one policy with each PutPermissionPolicy request.

The policy must include an Effect, Action and Principal.

Effect must specify Allow.

The Action in the policy must be waf:UpdateWebACL, waf-regional:UpdateWebACL, waf:GetRuleGroup and waf-regional:GetRuleGroup . Any extra or wildcard actions in the policy will be rejected.

The policy cannot include a Resource parameter.

The ARN in the request must be a valid WAF RuleGroup ARN and the RuleGroup must exist in the same region.

The user making the request must be the owner of the RuleGroup.

Your policy must be composed using IAM Policy version 2012-10-17.

For more information, see IAM Policies.

An example of a valid policy parameter is shown in the Examples section below.

Creating a Request

data PutPermissionPolicy

See: newPutPermissionPolicy smart constructor.

Constructors

PutPermissionPolicy'

Fields

resourceArn :: Text

The Amazon Resource Name (ARN) of the RuleGroup to which you want to attach the policy.

policy :: Text

The policy to attach to the specified RuleGroup.

newPutPermissionPolicy

Arguments

:: Text
$sel:resourceArn:PutPermissionPolicy'

-> Text
$sel:policy:PutPermissionPolicy'

-> PutPermissionPolicy

Create a value of PutPermissionPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:PutPermissionPolicy', putPermissionPolicy_resourceArn - The Amazon Resource Name (ARN) of the RuleGroup to which you want to attach the policy.

$sel:policy:PutPermissionPolicy', putPermissionPolicy_policy - The policy to attach to the specified RuleGroup.

Request Lenses

putPermissionPolicy_resourceArn :: Lens' PutPermissionPolicy Text

The Amazon Resource Name (ARN) of the RuleGroup to which you want to attach the policy.

putPermissionPolicy_policy :: Lens' PutPermissionPolicy Text

The policy to attach to the specified RuleGroup.

Destructuring the Response

data PutPermissionPolicyResponse

See: newPutPermissionPolicyResponse smart constructor.

Constructors

PutPermissionPolicyResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutPermissionPolicyResponse

Arguments

:: Int
$sel:httpStatus:PutPermissionPolicyResponse'

-> PutPermissionPolicyResponse

Create a value of PutPermissionPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutPermissionPolicyResponse', putPermissionPolicyResponse_httpStatus - The response's http status code.

Response Lenses

putPermissionPolicyResponse_httpStatus :: Lens' PutPermissionPolicyResponse Int

The response's http status code.