Network.AWS.WAF.Types.RegexMatchSet

Description

Documentation

data RegexMatchSet

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

In a GetRegexMatchSet request, RegexMatchSet is a complex type that contains the RegexMatchSetId and Name of a RegexMatchSet, and the values that you specified when you updated the RegexMatchSet.

The values are contained in a RegexMatchTuple object, which specify the parts of web requests that you want AWS WAF to inspect and the values that you want AWS WAF to search for. If a RegexMatchSet contains more than one RegexMatchTuple object, a request needs to match the settings in only one ByteMatchTuple to be considered a match.

See: newRegexMatchSet smart constructor.

Constructors

RegexMatchSet'

Fields

regexMatchSetId :: Maybe Text

The RegexMatchSetId for a RegexMatchSet. You use RegexMatchSetId to get information about a RegexMatchSet (see GetRegexMatchSet), update a RegexMatchSet (see UpdateRegexMatchSet), insert a RegexMatchSet into a Rule or delete one from a Rule (see UpdateRule), and delete a RegexMatchSet from AWS WAF (see DeleteRegexMatchSet).

RegexMatchSetId is returned by CreateRegexMatchSet and by ListRegexMatchSets.

regexMatchTuples :: Maybe [RegexMatchTuple]

Contains an array of RegexMatchTuple objects. Each RegexMatchTuple object contains:

The part of a web request that you want AWS WAF to inspect, such as a query string or the value of the User-Agent header.

The identifier of the pattern (a regular expression) that you want AWS WAF to look for. For more information, see RegexPatternSet.

Whether to perform any conversions on the request, such as converting it to lowercase, before inspecting it for the specified string.

name :: Maybe Text

A friendly name or description of the RegexMatchSet. You can't change Name after you create a RegexMatchSet.

newRegexMatchSet :: RegexMatchSet

Create a value of RegexMatchSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:regexMatchSetId:RegexMatchSet', regexMatchSet_regexMatchSetId - The RegexMatchSetId for a RegexMatchSet. You use RegexMatchSetId to get information about a RegexMatchSet (see GetRegexMatchSet), update a RegexMatchSet (see UpdateRegexMatchSet), insert a RegexMatchSet into a Rule or delete one from a Rule (see UpdateRule), and delete a RegexMatchSet from AWS WAF (see DeleteRegexMatchSet).

RegexMatchSetId is returned by CreateRegexMatchSet and by ListRegexMatchSets.

$sel:regexMatchTuples:RegexMatchSet', regexMatchSet_regexMatchTuples - Contains an array of RegexMatchTuple objects. Each RegexMatchTuple object contains:

The part of a web request that you want AWS WAF to inspect, such as a query string or the value of the User-Agent header.

The identifier of the pattern (a regular expression) that you want AWS WAF to look for. For more information, see RegexPatternSet.

Whether to perform any conversions on the request, such as converting it to lowercase, before inspecting it for the specified string.

$sel:name:RegexMatchSet', regexMatchSet_name - A friendly name or description of the RegexMatchSet. You can't change Name after you create a RegexMatchSet.

regexMatchSet_regexMatchSetId :: Lens' RegexMatchSet (Maybe Text)

The RegexMatchSetId for a RegexMatchSet. You use RegexMatchSetId to get information about a RegexMatchSet (see GetRegexMatchSet), update a RegexMatchSet (see UpdateRegexMatchSet), insert a RegexMatchSet into a Rule or delete one from a Rule (see UpdateRule), and delete a RegexMatchSet from AWS WAF (see DeleteRegexMatchSet).

RegexMatchSetId is returned by CreateRegexMatchSet and by ListRegexMatchSets.

regexMatchSet_regexMatchTuples :: Lens' RegexMatchSet (Maybe [RegexMatchTuple])

Contains an array of RegexMatchTuple objects. Each RegexMatchTuple object contains:

The part of a web request that you want AWS WAF to inspect, such as a query string or the value of the User-Agent header.

The identifier of the pattern (a regular expression) that you want AWS WAF to look for. For more information, see RegexPatternSet.

Whether to perform any conversions on the request, such as converting it to lowercase, before inspecting it for the specified string.

regexMatchSet_name :: Lens' RegexMatchSet (Maybe Text)

A friendly name or description of the RegexMatchSet. You can't change Name after you create a RegexMatchSet.