Network.AWS.WAF.Types.RegexPatternSetSummary

Description

Documentation

data RegexPatternSetSummary

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Returned by ListRegexPatternSets. Each RegexPatternSetSummary object includes the Name and RegexPatternSetId for one RegexPatternSet.

See: newRegexPatternSetSummary smart constructor.

Constructors

RegexPatternSetSummary'

Fields

regexPatternSetId :: Text

The RegexPatternSetId for a RegexPatternSet. You use RegexPatternSetId to get information about a RegexPatternSet, update a RegexPatternSet, remove a RegexPatternSet from a RegexMatchSet, and delete a RegexPatternSet from AWS WAF.

RegexPatternSetId is returned by CreateRegexPatternSet and by ListRegexPatternSets.

name :: Text

A friendly name or description of the RegexPatternSet. You can't change Name after you create a RegexPatternSet.

newRegexPatternSetSummary

Arguments

:: Text
$sel:regexPatternSetId:RegexPatternSetSummary'

-> Text
$sel:name:RegexPatternSetSummary'

-> RegexPatternSetSummary

Create a value of RegexPatternSetSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:regexPatternSetId:RegexPatternSetSummary', regexPatternSetSummary_regexPatternSetId - The RegexPatternSetId for a RegexPatternSet. You use RegexPatternSetId to get information about a RegexPatternSet, update a RegexPatternSet, remove a RegexPatternSet from a RegexMatchSet, and delete a RegexPatternSet from AWS WAF.

RegexPatternSetId is returned by CreateRegexPatternSet and by ListRegexPatternSets.

$sel:name:RegexPatternSetSummary', regexPatternSetSummary_name - A friendly name or description of the RegexPatternSet. You can't change Name after you create a RegexPatternSet.

regexPatternSetSummary_regexPatternSetId :: Lens' RegexPatternSetSummary Text

The RegexPatternSetId for a RegexPatternSet. You use RegexPatternSetId to get information about a RegexPatternSet, update a RegexPatternSet, remove a RegexPatternSet from a RegexMatchSet, and delete a RegexPatternSet from AWS WAF.

RegexPatternSetId is returned by CreateRegexPatternSet and by ListRegexPatternSets.

regexPatternSetSummary_name :: Lens' RegexPatternSetSummary Text

A friendly name or description of the RegexPatternSet. You can't change Name after you create a RegexPatternSet.