Network.AWS.WAF.UpdateByteMatchSet

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Inserts or deletes ByteMatchTuple objects (filters) in a ByteMatchSet. For each ByteMatchTuple object, you specify the following values:

Whether to insert or delete the object from the array. If you want to change a ByteMatchSetUpdate object, you delete the existing object and add a new one.

The part of a web request that you want AWS WAF to inspect, such as a query string or the value of the User-Agent header.

The bytes (typically a string that corresponds with ASCII characters) that you want AWS WAF to look for. For more information, including how you specify the values for the AWS WAF API and the AWS CLI or SDKs, see TargetString in the ByteMatchTuple data type.

Where to look, such as at the beginning or the end of a query string.

Whether to perform any conversions on the request, such as converting it to lowercase, before inspecting it for the specified string.

For example, you can add a ByteMatchSetUpdate object that matches web requests in which User-Agent headers contain the string BadBot. You can then configure AWS WAF to block those requests.

To create and configure a ByteMatchSet, perform the following steps:

Create a ByteMatchSet. For more information, see CreateByteMatchSet.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateByteMatchSet request.

Submit an UpdateByteMatchSet request to specify the part of the request that you want AWS WAF to inspect (for example, the header or the URI) and the value that you want AWS WAF to watch for.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide.

Creating a Request

data UpdateByteMatchSet

See: newUpdateByteMatchSet smart constructor.

Constructors

UpdateByteMatchSet'

Fields

byteMatchSetId :: Text

The ByteMatchSetId of the ByteMatchSet that you want to update. ByteMatchSetId is returned by CreateByteMatchSet and by ListByteMatchSets.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

updates :: NonEmpty ByteMatchSetUpdate

An array of ByteMatchSetUpdate objects that you want to insert into or delete from a ByteMatchSet. For more information, see the applicable data types:

ByteMatchSetUpdate: Contains Action and ByteMatchTuple

ByteMatchTuple: Contains FieldToMatch, PositionalConstraint, TargetString, and TextTransformation

FieldToMatch: Contains Data and Type

newUpdateByteMatchSet

Arguments

:: Text
$sel:byteMatchSetId:UpdateByteMatchSet'

-> Text
$sel:changeToken:UpdateByteMatchSet'

-> NonEmpty ByteMatchSetUpdate
$sel:updates:UpdateByteMatchSet'

-> UpdateByteMatchSet

Create a value of UpdateByteMatchSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:byteMatchSetId:UpdateByteMatchSet', updateByteMatchSet_byteMatchSetId - The ByteMatchSetId of the ByteMatchSet that you want to update. ByteMatchSetId is returned by CreateByteMatchSet and by ListByteMatchSets.

$sel:changeToken:UpdateByteMatchSet', updateByteMatchSet_changeToken - The value returned by the most recent call to GetChangeToken.

$sel:updates:UpdateByteMatchSet', updateByteMatchSet_updates - An array of ByteMatchSetUpdate objects that you want to insert into or delete from a ByteMatchSet. For more information, see the applicable data types:

ByteMatchSetUpdate: Contains Action and ByteMatchTuple

ByteMatchTuple: Contains FieldToMatch, PositionalConstraint, TargetString, and TextTransformation

FieldToMatch: Contains Data and Type

Request Lenses

updateByteMatchSet_byteMatchSetId :: Lens' UpdateByteMatchSet Text

The ByteMatchSetId of the ByteMatchSet that you want to update. ByteMatchSetId is returned by CreateByteMatchSet and by ListByteMatchSets.

updateByteMatchSet_changeToken :: Lens' UpdateByteMatchSet Text

The value returned by the most recent call to GetChangeToken.

updateByteMatchSet_updates :: Lens' UpdateByteMatchSet (NonEmpty ByteMatchSetUpdate)

An array of ByteMatchSetUpdate objects that you want to insert into or delete from a ByteMatchSet. For more information, see the applicable data types:

ByteMatchSetUpdate: Contains Action and ByteMatchTuple

ByteMatchTuple: Contains FieldToMatch, PositionalConstraint, TargetString, and TextTransformation

FieldToMatch: Contains Data and Type

Destructuring the Response

data UpdateByteMatchSetResponse

See: newUpdateByteMatchSetResponse smart constructor.

Constructors

UpdateByteMatchSetResponse'

Fields

changeToken :: Maybe Text

The ChangeToken that you used to submit the UpdateByteMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newUpdateByteMatchSetResponse

Arguments

:: Int
$sel:httpStatus:UpdateByteMatchSetResponse'

-> UpdateByteMatchSetResponse

Create a value of UpdateByteMatchSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeToken:UpdateByteMatchSet', updateByteMatchSetResponse_changeToken - The ChangeToken that you used to submit the UpdateByteMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:UpdateByteMatchSetResponse', updateByteMatchSetResponse_httpStatus - The response's http status code.

Response Lenses

updateByteMatchSetResponse_changeToken :: Lens' UpdateByteMatchSetResponse (Maybe Text)

The ChangeToken that you used to submit the UpdateByteMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

updateByteMatchSetResponse_httpStatus :: Lens' UpdateByteMatchSetResponse Int

The response's http status code.