Network.AWS.WAF.UpdateGeoMatchSet

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Inserts or deletes GeoMatchConstraint objects in an GeoMatchSet. For each GeoMatchConstraint object, you specify the following values:

Whether to insert or delete the object from the array. If you want to change an GeoMatchConstraint object, you delete the existing object and add a new one.

The Type. The only valid value for Type is Country.

The Value, which is a two character code for the country to add to the GeoMatchConstraint object. Valid codes are listed in GeoMatchConstraint$Value.

To create and configure an GeoMatchSet, perform the following steps:

Submit a CreateGeoMatchSet request.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateGeoMatchSet request.

Submit an UpdateGeoMatchSet request to specify the country that you want AWS WAF to watch for.

When you update an GeoMatchSet, you specify the country that you want to add and/or the country that you want to delete. If you want to change a country, you delete the existing country and add the new one.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide.

Creating a Request

data UpdateGeoMatchSet

See: newUpdateGeoMatchSet smart constructor.

Constructors

UpdateGeoMatchSet'

Fields

geoMatchSetId :: Text

The GeoMatchSetId of the GeoMatchSet that you want to update. GeoMatchSetId is returned by CreateGeoMatchSet and by ListGeoMatchSets.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

updates :: NonEmpty GeoMatchSetUpdate

An array of GeoMatchSetUpdate objects that you want to insert into or delete from an GeoMatchSet. For more information, see the applicable data types:

GeoMatchSetUpdate: Contains Action and GeoMatchConstraint

GeoMatchConstraint: Contains Type and Value

You can have only one Type and Value per GeoMatchConstraint. To add multiple countries, include multiple GeoMatchSetUpdate objects in your request.

newUpdateGeoMatchSet

Arguments

:: Text
$sel:geoMatchSetId:UpdateGeoMatchSet'

-> Text
$sel:changeToken:UpdateGeoMatchSet'

-> NonEmpty GeoMatchSetUpdate
$sel:updates:UpdateGeoMatchSet'

-> UpdateGeoMatchSet

Create a value of UpdateGeoMatchSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:geoMatchSetId:UpdateGeoMatchSet', updateGeoMatchSet_geoMatchSetId - The GeoMatchSetId of the GeoMatchSet that you want to update. GeoMatchSetId is returned by CreateGeoMatchSet and by ListGeoMatchSets.

$sel:changeToken:UpdateGeoMatchSet', updateGeoMatchSet_changeToken - The value returned by the most recent call to GetChangeToken.

$sel:updates:UpdateGeoMatchSet', updateGeoMatchSet_updates - An array of GeoMatchSetUpdate objects that you want to insert into or delete from an GeoMatchSet. For more information, see the applicable data types:

GeoMatchSetUpdate: Contains Action and GeoMatchConstraint

GeoMatchConstraint: Contains Type and Value

You can have only one Type and Value per GeoMatchConstraint. To add multiple countries, include multiple GeoMatchSetUpdate objects in your request.

Request Lenses

updateGeoMatchSet_geoMatchSetId :: Lens' UpdateGeoMatchSet Text

The GeoMatchSetId of the GeoMatchSet that you want to update. GeoMatchSetId is returned by CreateGeoMatchSet and by ListGeoMatchSets.

updateGeoMatchSet_changeToken :: Lens' UpdateGeoMatchSet Text

The value returned by the most recent call to GetChangeToken.

updateGeoMatchSet_updates :: Lens' UpdateGeoMatchSet (NonEmpty GeoMatchSetUpdate)

An array of GeoMatchSetUpdate objects that you want to insert into or delete from an GeoMatchSet. For more information, see the applicable data types:

GeoMatchSetUpdate: Contains Action and GeoMatchConstraint

GeoMatchConstraint: Contains Type and Value

You can have only one Type and Value per GeoMatchConstraint. To add multiple countries, include multiple GeoMatchSetUpdate objects in your request.

Destructuring the Response

data UpdateGeoMatchSetResponse

See: newUpdateGeoMatchSetResponse smart constructor.

Constructors

UpdateGeoMatchSetResponse'

Fields

changeToken :: Maybe Text

The ChangeToken that you used to submit the UpdateGeoMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newUpdateGeoMatchSetResponse

Arguments

:: Int
$sel:httpStatus:UpdateGeoMatchSetResponse'

-> UpdateGeoMatchSetResponse

Create a value of UpdateGeoMatchSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeToken:UpdateGeoMatchSet', updateGeoMatchSetResponse_changeToken - The ChangeToken that you used to submit the UpdateGeoMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:UpdateGeoMatchSetResponse', updateGeoMatchSetResponse_httpStatus - The response's http status code.

Response Lenses

updateGeoMatchSetResponse_changeToken :: Lens' UpdateGeoMatchSetResponse (Maybe Text)

The ChangeToken that you used to submit the UpdateGeoMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

updateGeoMatchSetResponse_httpStatus :: Lens' UpdateGeoMatchSetResponse Int

The response's http status code.