Network.AWS.WAF.UpdateSizeConstraintSet

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Inserts or deletes SizeConstraint objects (filters) in a SizeConstraintSet. For each SizeConstraint object, you specify the following values:

Whether to insert or delete the object from the array. If you want to change a SizeConstraintSetUpdate object, you delete the existing object and add a new one.

The part of a web request that you want AWS WAF to evaluate, such as the length of a query string or the length of the User-Agent header.

Whether to perform any transformations on the request, such as converting it to lowercase, before checking its length. Note that transformations of the request body are not supported because the AWS resource forwards only the first 8192 bytes of your request to AWS WAF.

You can only specify a single type of TextTransformation.

A ComparisonOperator used for evaluating the selected part of the request against the specified Size, such as equals, greater than, less than, and so on.

The length, in bytes, that you want AWS WAF to watch for in selected part of the request. The length is computed after applying the transformation.

For example, you can add a SizeConstraintSetUpdate object that matches web requests in which the length of the User-Agent header is greater than 100 bytes. You can then configure AWS WAF to block those requests.

To create and configure a SizeConstraintSet, perform the following steps:

Create a SizeConstraintSet. For more information, see CreateSizeConstraintSet.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateSizeConstraintSet request.

Submit an UpdateSizeConstraintSet request to specify the part of the request that you want AWS WAF to inspect (for example, the header or the URI) and the value that you want AWS WAF to watch for.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide.

Creating a Request

data UpdateSizeConstraintSet

See: newUpdateSizeConstraintSet smart constructor.

Constructors

UpdateSizeConstraintSet'

Fields

sizeConstraintSetId :: Text

The SizeConstraintSetId of the SizeConstraintSet that you want to update. SizeConstraintSetId is returned by CreateSizeConstraintSet and by ListSizeConstraintSets.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

updates :: NonEmpty SizeConstraintSetUpdate

An array of SizeConstraintSetUpdate objects that you want to insert into or delete from a SizeConstraintSet. For more information, see the applicable data types:

SizeConstraintSetUpdate: Contains Action and SizeConstraint

SizeConstraint: Contains FieldToMatch, TextTransformation, ComparisonOperator, and Size

FieldToMatch: Contains Data and Type

newUpdateSizeConstraintSet

Arguments

:: Text
$sel:sizeConstraintSetId:UpdateSizeConstraintSet'

-> Text
$sel:changeToken:UpdateSizeConstraintSet'

-> NonEmpty SizeConstraintSetUpdate
$sel:updates:UpdateSizeConstraintSet'

-> UpdateSizeConstraintSet

Create a value of UpdateSizeConstraintSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sizeConstraintSetId:UpdateSizeConstraintSet', updateSizeConstraintSet_sizeConstraintSetId - The SizeConstraintSetId of the SizeConstraintSet that you want to update. SizeConstraintSetId is returned by CreateSizeConstraintSet and by ListSizeConstraintSets.

$sel:changeToken:UpdateSizeConstraintSet', updateSizeConstraintSet_changeToken - The value returned by the most recent call to GetChangeToken.

$sel:updates:UpdateSizeConstraintSet', updateSizeConstraintSet_updates - An array of SizeConstraintSetUpdate objects that you want to insert into or delete from a SizeConstraintSet. For more information, see the applicable data types:

SizeConstraintSetUpdate: Contains Action and SizeConstraint

SizeConstraint: Contains FieldToMatch, TextTransformation, ComparisonOperator, and Size

FieldToMatch: Contains Data and Type

Request Lenses

updateSizeConstraintSet_sizeConstraintSetId :: Lens' UpdateSizeConstraintSet Text

The SizeConstraintSetId of the SizeConstraintSet that you want to update. SizeConstraintSetId is returned by CreateSizeConstraintSet and by ListSizeConstraintSets.

updateSizeConstraintSet_changeToken :: Lens' UpdateSizeConstraintSet Text

The value returned by the most recent call to GetChangeToken.

updateSizeConstraintSet_updates :: Lens' UpdateSizeConstraintSet (NonEmpty SizeConstraintSetUpdate)

An array of SizeConstraintSetUpdate objects that you want to insert into or delete from a SizeConstraintSet. For more information, see the applicable data types:

SizeConstraintSetUpdate: Contains Action and SizeConstraint

SizeConstraint: Contains FieldToMatch, TextTransformation, ComparisonOperator, and Size

FieldToMatch: Contains Data and Type

Destructuring the Response

data UpdateSizeConstraintSetResponse

See: newUpdateSizeConstraintSetResponse smart constructor.

Constructors

UpdateSizeConstraintSetResponse'

Fields

changeToken :: Maybe Text

The ChangeToken that you used to submit the UpdateSizeConstraintSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newUpdateSizeConstraintSetResponse

Arguments

:: Int
$sel:httpStatus:UpdateSizeConstraintSetResponse'

-> UpdateSizeConstraintSetResponse

Create a value of UpdateSizeConstraintSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeToken:UpdateSizeConstraintSet', updateSizeConstraintSetResponse_changeToken - The ChangeToken that you used to submit the UpdateSizeConstraintSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:UpdateSizeConstraintSetResponse', updateSizeConstraintSetResponse_httpStatus - The response's http status code.

Response Lenses

updateSizeConstraintSetResponse_changeToken :: Lens' UpdateSizeConstraintSetResponse (Maybe Text)

The ChangeToken that you used to submit the UpdateSizeConstraintSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

updateSizeConstraintSetResponse_httpStatus :: Lens' UpdateSizeConstraintSetResponse Int

The response's http status code.