Network.AWS.WAF.UpdateXssMatchSet

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Inserts or deletes XssMatchTuple objects (filters) in an XssMatchSet. For each XssMatchTuple object, you specify the following values:

Action: Whether to insert the object into or delete the object from the array. To change an XssMatchTuple, you delete the existing object and add a new one.

FieldToMatch: The part of web requests that you want AWS WAF to inspect and, if you want AWS WAF to inspect a header or custom query parameter, the name of the header or parameter.

TextTransformation: Which text transformation, if any, to perform on the web request before inspecting the request for cross-site scripting attacks.

You can only specify a single type of TextTransformation.

You use XssMatchSet objects to specify which CloudFront requests that you want to allow, block, or count. For example, if you're receiving requests that contain cross-site scripting attacks in the request body and you want to block the requests, you can create an XssMatchSet with the applicable settings, and then configure AWS WAF to block the requests.

To create and configure an XssMatchSet, perform the following steps:

Submit a CreateXssMatchSet request.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateIPSet request.

Submit an UpdateXssMatchSet request to specify the parts of web requests that you want AWS WAF to inspect for cross-site scripting attacks.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide.

Creating a Request

data UpdateXssMatchSet

A request to update an XssMatchSet.

See: newUpdateXssMatchSet smart constructor.

Constructors

UpdateXssMatchSet'

Fields

xssMatchSetId :: Text

The XssMatchSetId of the XssMatchSet that you want to update. XssMatchSetId is returned by CreateXssMatchSet and by ListXssMatchSets.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

updates :: NonEmpty XssMatchSetUpdate

An array of XssMatchSetUpdate objects that you want to insert into or delete from an XssMatchSet. For more information, see the applicable data types:

XssMatchSetUpdate: Contains Action and XssMatchTuple

XssMatchTuple: Contains FieldToMatch and TextTransformation

FieldToMatch: Contains Data and Type

newUpdateXssMatchSet

Arguments

:: Text
$sel:xssMatchSetId:UpdateXssMatchSet'

-> Text
$sel:changeToken:UpdateXssMatchSet'

-> NonEmpty XssMatchSetUpdate
$sel:updates:UpdateXssMatchSet'

-> UpdateXssMatchSet

Create a value of UpdateXssMatchSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:xssMatchSetId:UpdateXssMatchSet', updateXssMatchSet_xssMatchSetId - The XssMatchSetId of the XssMatchSet that you want to update. XssMatchSetId is returned by CreateXssMatchSet and by ListXssMatchSets.

$sel:changeToken:UpdateXssMatchSet', updateXssMatchSet_changeToken - The value returned by the most recent call to GetChangeToken.

$sel:updates:UpdateXssMatchSet', updateXssMatchSet_updates - An array of XssMatchSetUpdate objects that you want to insert into or delete from an XssMatchSet. For more information, see the applicable data types:

XssMatchSetUpdate: Contains Action and XssMatchTuple

XssMatchTuple: Contains FieldToMatch and TextTransformation

FieldToMatch: Contains Data and Type

Request Lenses

updateXssMatchSet_xssMatchSetId :: Lens' UpdateXssMatchSet Text

The XssMatchSetId of the XssMatchSet that you want to update. XssMatchSetId is returned by CreateXssMatchSet and by ListXssMatchSets.

updateXssMatchSet_changeToken :: Lens' UpdateXssMatchSet Text

The value returned by the most recent call to GetChangeToken.

updateXssMatchSet_updates :: Lens' UpdateXssMatchSet (NonEmpty XssMatchSetUpdate)

An array of XssMatchSetUpdate objects that you want to insert into or delete from an XssMatchSet. For more information, see the applicable data types:

XssMatchSetUpdate: Contains Action and XssMatchTuple

XssMatchTuple: Contains FieldToMatch and TextTransformation

FieldToMatch: Contains Data and Type

Destructuring the Response

data UpdateXssMatchSetResponse

The response to an UpdateXssMatchSets request.

See: newUpdateXssMatchSetResponse smart constructor.

Constructors

UpdateXssMatchSetResponse'

Fields

changeToken :: Maybe Text

The ChangeToken that you used to submit the UpdateXssMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newUpdateXssMatchSetResponse

Arguments

:: Int
$sel:httpStatus:UpdateXssMatchSetResponse'

-> UpdateXssMatchSetResponse

Create a value of UpdateXssMatchSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeToken:UpdateXssMatchSet', updateXssMatchSetResponse_changeToken - The ChangeToken that you used to submit the UpdateXssMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:UpdateXssMatchSetResponse', updateXssMatchSetResponse_httpStatus - The response's http status code.

Response Lenses

updateXssMatchSetResponse_changeToken :: Lens' UpdateXssMatchSetResponse (Maybe Text)

The ChangeToken that you used to submit the UpdateXssMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

updateXssMatchSetResponse_httpStatus :: Lens' UpdateXssMatchSetResponse Int

The response's http status code.