Network.AWS.WAFRegional.CreateRegexPatternSet

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Creates a RegexPatternSet. You then use UpdateRegexPatternSet to specify the regular expression (regex) pattern that you want AWS WAF to search for, such as B[a@]dB[o0]t. You can then configure AWS WAF to reject those requests.

To create and configure a RegexPatternSet, perform the following steps:

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a CreateRegexPatternSet request.

Submit a CreateRegexPatternSet request.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateRegexPatternSet request.

Submit an UpdateRegexPatternSet request to specify the string that you want AWS WAF to watch for.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide.

Creating a Request

data CreateRegexPatternSet

See: newCreateRegexPatternSet smart constructor.

Constructors

CreateRegexPatternSet'

Fields

name :: Text

A friendly name or description of the RegexPatternSet. You can't change Name after you create a RegexPatternSet.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

newCreateRegexPatternSet

Arguments

:: Text
$sel:name:CreateRegexPatternSet'

-> Text
$sel:changeToken:CreateRegexPatternSet'

-> CreateRegexPatternSet

Create a value of CreateRegexPatternSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreateRegexPatternSet', createRegexPatternSet_name - A friendly name or description of the RegexPatternSet. You can't change Name after you create a RegexPatternSet.

$sel:changeToken:CreateRegexPatternSet', createRegexPatternSet_changeToken - The value returned by the most recent call to GetChangeToken.

Request Lenses

createRegexPatternSet_name :: Lens' CreateRegexPatternSet Text

A friendly name or description of the RegexPatternSet. You can't change Name after you create a RegexPatternSet.

createRegexPatternSet_changeToken :: Lens' CreateRegexPatternSet Text

The value returned by the most recent call to GetChangeToken.

Destructuring the Response

data CreateRegexPatternSetResponse

See: newCreateRegexPatternSetResponse smart constructor.

Constructors

CreateRegexPatternSetResponse'

Fields

regexPatternSet :: Maybe RegexPatternSet

A RegexPatternSet that contains no objects.

changeToken :: Maybe Text

The ChangeToken that you used to submit the CreateRegexPatternSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newCreateRegexPatternSetResponse

Arguments

:: Int
$sel:httpStatus:CreateRegexPatternSetResponse'

-> CreateRegexPatternSetResponse

Create a value of CreateRegexPatternSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:regexPatternSet:CreateRegexPatternSetResponse', createRegexPatternSetResponse_regexPatternSet - A RegexPatternSet that contains no objects.

$sel:changeToken:CreateRegexPatternSet', createRegexPatternSetResponse_changeToken - The ChangeToken that you used to submit the CreateRegexPatternSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:CreateRegexPatternSetResponse', createRegexPatternSetResponse_httpStatus - The response's http status code.

Response Lenses

createRegexPatternSetResponse_regexPatternSet :: Lens' CreateRegexPatternSetResponse (Maybe RegexPatternSet)

A RegexPatternSet that contains no objects.

createRegexPatternSetResponse_changeToken :: Lens' CreateRegexPatternSetResponse (Maybe Text)

The ChangeToken that you used to submit the CreateRegexPatternSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

createRegexPatternSetResponse_httpStatus :: Lens' CreateRegexPatternSetResponse Int

The response's http status code.