Network.AWS.WAFRegional.CreateRuleGroup

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Creates a RuleGroup. A rule group is a collection of predefined rules that you add to a web ACL. You use UpdateRuleGroup to add rules to the rule group.

Rule groups are subject to the following limits:

Three rule groups per account. You can request an increase to this limit by contacting customer support.

One rule group per web ACL.

Ten rules per rule group.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide.

Creating a Request

data CreateRuleGroup

See: newCreateRuleGroup smart constructor.

Constructors

CreateRuleGroup'

Fields

tags :: Maybe (NonEmpty Tag)

name :: Text

A friendly name or description of the RuleGroup. You can't change Name after you create a RuleGroup.

metricName :: Text

A friendly name or description for the metrics for this RuleGroup. The name can contain only alphanumeric characters (A-Z, a-z, 0-9), with maximum length 128 and minimum length one. It can't contain whitespace or metric names reserved for AWS WAF, including "All" and "Default_Action." You can't change the name of the metric after you create the RuleGroup.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

newCreateRuleGroup

Arguments

:: Text
$sel:name:CreateRuleGroup'

-> Text
$sel:metricName:CreateRuleGroup'

-> Text
$sel:changeToken:CreateRuleGroup'

-> CreateRuleGroup

Create a value of CreateRuleGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateRuleGroup', createRuleGroup_tags -

$sel:name:CreateRuleGroup', createRuleGroup_name - A friendly name or description of the RuleGroup. You can't change Name after you create a RuleGroup.

$sel:metricName:CreateRuleGroup', createRuleGroup_metricName - A friendly name or description for the metrics for this RuleGroup. The name can contain only alphanumeric characters (A-Z, a-z, 0-9), with maximum length 128 and minimum length one. It can't contain whitespace or metric names reserved for AWS WAF, including "All" and "Default_Action." You can't change the name of the metric after you create the RuleGroup.

$sel:changeToken:CreateRuleGroup', createRuleGroup_changeToken - The value returned by the most recent call to GetChangeToken.

Request Lenses

createRuleGroup_tags :: Lens' CreateRuleGroup (Maybe (NonEmpty Tag))

createRuleGroup_name :: Lens' CreateRuleGroup Text

A friendly name or description of the RuleGroup. You can't change Name after you create a RuleGroup.

createRuleGroup_metricName :: Lens' CreateRuleGroup Text

A friendly name or description for the metrics for this RuleGroup. The name can contain only alphanumeric characters (A-Z, a-z, 0-9), with maximum length 128 and minimum length one. It can't contain whitespace or metric names reserved for AWS WAF, including "All" and "Default_Action." You can't change the name of the metric after you create the RuleGroup.

createRuleGroup_changeToken :: Lens' CreateRuleGroup Text

The value returned by the most recent call to GetChangeToken.

Destructuring the Response

data CreateRuleGroupResponse

See: newCreateRuleGroupResponse smart constructor.

Constructors

CreateRuleGroupResponse'

Fields

ruleGroup :: Maybe RuleGroup

An empty RuleGroup.

changeToken :: Maybe Text

The ChangeToken that you used to submit the CreateRuleGroup request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newCreateRuleGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateRuleGroupResponse'

-> CreateRuleGroupResponse

Create a value of CreateRuleGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleGroup:CreateRuleGroupResponse', createRuleGroupResponse_ruleGroup - An empty RuleGroup.

$sel:changeToken:CreateRuleGroup', createRuleGroupResponse_changeToken - The ChangeToken that you used to submit the CreateRuleGroup request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:CreateRuleGroupResponse', createRuleGroupResponse_httpStatus - The response's http status code.

Response Lenses

createRuleGroupResponse_ruleGroup :: Lens' CreateRuleGroupResponse (Maybe RuleGroup)

An empty RuleGroup.

createRuleGroupResponse_changeToken :: Lens' CreateRuleGroupResponse (Maybe Text)

The ChangeToken that you used to submit the CreateRuleGroup request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

createRuleGroupResponse_httpStatus :: Lens' CreateRuleGroupResponse Int

The response's http status code.