Network.AWS.WAFRegional.CreateWebACLMigrationStack

Description

Creates an AWS CloudFormation WAFV2 template for the specified web ACL in the specified Amazon S3 bucket. Then, in CloudFormation, you create a stack from the template, to create the web ACL and its resources in AWS WAFV2. Use this to migrate your AWS WAF Classic web ACL to the latest version of AWS WAF.

This is part of a larger migration procedure for web ACLs from AWS WAF Classic to the latest version of AWS WAF. For the full procedure, including caveats and manual steps to complete the migration and switch over to the new web ACL, see Migrating your AWS WAF Classic resources to AWS WAF in the AWS WAF Developer Guide.

Creating a Request

data CreateWebACLMigrationStack

See: newCreateWebACLMigrationStack smart constructor.

Constructors

CreateWebACLMigrationStack'

Fields

webACLId :: Text

The UUID of the WAF Classic web ACL that you want to migrate to WAF v2.

s3BucketName :: Text

The name of the Amazon S3 bucket to store the CloudFormation template in. The S3 bucket must be configured as follows for the migration:

The bucket name must start with aws-waf-migration-. For example, aws-waf-migration-my-web-acl.

The bucket must be in the Region where you are deploying the template. For example, for a web ACL in us-west-2, you must use an Amazon S3 bucket in us-west-2 and you must deploy the template stack to us-west-2.

The bucket policies must permit the migration process to write data. For listings of the bucket policies, see the Examples section.

ignoreUnsupportedType :: Bool

Indicates whether to exclude entities that can't be migrated or to stop the migration. Set this to true to ignore unsupported entities in the web ACL during the migration. Otherwise, if AWS WAF encounters unsupported entities, it stops the process and throws an exception.

newCreateWebACLMigrationStack

Arguments

:: Text
$sel:webACLId:CreateWebACLMigrationStack'

-> Text
$sel:s3BucketName:CreateWebACLMigrationStack'

-> Bool
$sel:ignoreUnsupportedType:CreateWebACLMigrationStack'

-> CreateWebACLMigrationStack

Create a value of CreateWebACLMigrationStack with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:webACLId:CreateWebACLMigrationStack', createWebACLMigrationStack_webACLId - The UUID of the WAF Classic web ACL that you want to migrate to WAF v2.

$sel:s3BucketName:CreateWebACLMigrationStack', createWebACLMigrationStack_s3BucketName - The name of the Amazon S3 bucket to store the CloudFormation template in. The S3 bucket must be configured as follows for the migration:

The bucket name must start with aws-waf-migration-. For example, aws-waf-migration-my-web-acl.

The bucket must be in the Region where you are deploying the template. For example, for a web ACL in us-west-2, you must use an Amazon S3 bucket in us-west-2 and you must deploy the template stack to us-west-2.

The bucket policies must permit the migration process to write data. For listings of the bucket policies, see the Examples section.

$sel:ignoreUnsupportedType:CreateWebACLMigrationStack', createWebACLMigrationStack_ignoreUnsupportedType - Indicates whether to exclude entities that can't be migrated or to stop the migration. Set this to true to ignore unsupported entities in the web ACL during the migration. Otherwise, if AWS WAF encounters unsupported entities, it stops the process and throws an exception.

Request Lenses

createWebACLMigrationStack_webACLId :: Lens' CreateWebACLMigrationStack Text

The UUID of the WAF Classic web ACL that you want to migrate to WAF v2.

createWebACLMigrationStack_s3BucketName :: Lens' CreateWebACLMigrationStack Text

The name of the Amazon S3 bucket to store the CloudFormation template in. The S3 bucket must be configured as follows for the migration:

The bucket name must start with aws-waf-migration-. For example, aws-waf-migration-my-web-acl.

The bucket must be in the Region where you are deploying the template. For example, for a web ACL in us-west-2, you must use an Amazon S3 bucket in us-west-2 and you must deploy the template stack to us-west-2.

The bucket policies must permit the migration process to write data. For listings of the bucket policies, see the Examples section.

createWebACLMigrationStack_ignoreUnsupportedType :: Lens' CreateWebACLMigrationStack Bool

Indicates whether to exclude entities that can't be migrated or to stop the migration. Set this to true to ignore unsupported entities in the web ACL during the migration. Otherwise, if AWS WAF encounters unsupported entities, it stops the process and throws an exception.

Destructuring the Response

data CreateWebACLMigrationStackResponse

See: newCreateWebACLMigrationStackResponse smart constructor.

Constructors

CreateWebACLMigrationStackResponse'

Fields

httpStatus :: Int

The response's http status code.

s3ObjectUrl :: Text

The URL of the template created in Amazon S3.

newCreateWebACLMigrationStackResponse

Arguments

:: Int
$sel:httpStatus:CreateWebACLMigrationStackResponse'

-> Text
$sel:s3ObjectUrl:CreateWebACLMigrationStackResponse'

-> CreateWebACLMigrationStackResponse

Create a value of CreateWebACLMigrationStackResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateWebACLMigrationStackResponse', createWebACLMigrationStackResponse_httpStatus - The response's http status code.

$sel:s3ObjectUrl:CreateWebACLMigrationStackResponse', createWebACLMigrationStackResponse_s3ObjectUrl - The URL of the template created in Amazon S3.

Response Lenses

createWebACLMigrationStackResponse_httpStatus :: Lens' CreateWebACLMigrationStackResponse Int

The response's http status code.

createWebACLMigrationStackResponse_s3ObjectUrl :: Lens' CreateWebACLMigrationStackResponse Text

The URL of the template created in Amazon S3.