Network.AWS.WAFRegional.DeleteByteMatchSet

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Permanently deletes a ByteMatchSet. You can't delete a ByteMatchSet if it's still used in any Rules or if it still includes any ByteMatchTuple objects (any filters).

If you just want to remove a ByteMatchSet from a Rule, use UpdateRule.

To permanently delete a ByteMatchSet, perform the following steps:

Update the ByteMatchSet to remove filters, if any. For more information, see UpdateByteMatchSet.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a DeleteByteMatchSet request.

Submit a DeleteByteMatchSet request.

Creating a Request

data DeleteByteMatchSet

See: newDeleteByteMatchSet smart constructor.

Constructors

DeleteByteMatchSet'

Fields

byteMatchSetId :: Text

The ByteMatchSetId of the ByteMatchSet that you want to delete. ByteMatchSetId is returned by CreateByteMatchSet and by ListByteMatchSets.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

newDeleteByteMatchSet

Arguments

:: Text
$sel:byteMatchSetId:DeleteByteMatchSet'

-> Text
$sel:changeToken:DeleteByteMatchSet'

-> DeleteByteMatchSet

Create a value of DeleteByteMatchSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:byteMatchSetId:DeleteByteMatchSet', deleteByteMatchSet_byteMatchSetId - The ByteMatchSetId of the ByteMatchSet that you want to delete. ByteMatchSetId is returned by CreateByteMatchSet and by ListByteMatchSets.

$sel:changeToken:DeleteByteMatchSet', deleteByteMatchSet_changeToken - The value returned by the most recent call to GetChangeToken.

Request Lenses

deleteByteMatchSet_byteMatchSetId :: Lens' DeleteByteMatchSet Text

The ByteMatchSetId of the ByteMatchSet that you want to delete. ByteMatchSetId is returned by CreateByteMatchSet and by ListByteMatchSets.

deleteByteMatchSet_changeToken :: Lens' DeleteByteMatchSet Text

The value returned by the most recent call to GetChangeToken.

Destructuring the Response

data DeleteByteMatchSetResponse

See: newDeleteByteMatchSetResponse smart constructor.

Constructors

DeleteByteMatchSetResponse'

Fields

changeToken :: Maybe Text

The ChangeToken that you used to submit the DeleteByteMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newDeleteByteMatchSetResponse

Arguments

:: Int
$sel:httpStatus:DeleteByteMatchSetResponse'

-> DeleteByteMatchSetResponse

Create a value of DeleteByteMatchSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeToken:DeleteByteMatchSet', deleteByteMatchSetResponse_changeToken - The ChangeToken that you used to submit the DeleteByteMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:DeleteByteMatchSetResponse', deleteByteMatchSetResponse_httpStatus - The response's http status code.

Response Lenses

deleteByteMatchSetResponse_changeToken :: Lens' DeleteByteMatchSetResponse (Maybe Text)

The ChangeToken that you used to submit the DeleteByteMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

deleteByteMatchSetResponse_httpStatus :: Lens' DeleteByteMatchSetResponse Int

The response's http status code.