Network.AWS.WAFRegional.DeleteGeoMatchSet

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Permanently deletes a GeoMatchSet. You can't delete a GeoMatchSet if it's still used in any Rules or if it still includes any countries.

If you just want to remove a GeoMatchSet from a Rule, use UpdateRule.

To permanently delete a GeoMatchSet from AWS WAF, perform the following steps:

Update the GeoMatchSet to remove any countries. For more information, see UpdateGeoMatchSet.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a DeleteGeoMatchSet request.

Submit a DeleteGeoMatchSet request.

Creating a Request

data DeleteGeoMatchSet

See: newDeleteGeoMatchSet smart constructor.

Constructors

DeleteGeoMatchSet'

Fields

geoMatchSetId :: Text

The GeoMatchSetID of the GeoMatchSet that you want to delete. GeoMatchSetId is returned by CreateGeoMatchSet and by ListGeoMatchSets.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

newDeleteGeoMatchSet

Arguments

:: Text
$sel:geoMatchSetId:DeleteGeoMatchSet'

-> Text
$sel:changeToken:DeleteGeoMatchSet'

-> DeleteGeoMatchSet

Create a value of DeleteGeoMatchSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:geoMatchSetId:DeleteGeoMatchSet', deleteGeoMatchSet_geoMatchSetId - The GeoMatchSetID of the GeoMatchSet that you want to delete. GeoMatchSetId is returned by CreateGeoMatchSet and by ListGeoMatchSets.

$sel:changeToken:DeleteGeoMatchSet', deleteGeoMatchSet_changeToken - The value returned by the most recent call to GetChangeToken.

Request Lenses

deleteGeoMatchSet_geoMatchSetId :: Lens' DeleteGeoMatchSet Text

The GeoMatchSetID of the GeoMatchSet that you want to delete. GeoMatchSetId is returned by CreateGeoMatchSet and by ListGeoMatchSets.

deleteGeoMatchSet_changeToken :: Lens' DeleteGeoMatchSet Text

The value returned by the most recent call to GetChangeToken.

Destructuring the Response

data DeleteGeoMatchSetResponse

See: newDeleteGeoMatchSetResponse smart constructor.

Constructors

DeleteGeoMatchSetResponse'

Fields

changeToken :: Maybe Text

The ChangeToken that you used to submit the DeleteGeoMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newDeleteGeoMatchSetResponse

Arguments

:: Int
$sel:httpStatus:DeleteGeoMatchSetResponse'

-> DeleteGeoMatchSetResponse

Create a value of DeleteGeoMatchSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeToken:DeleteGeoMatchSet', deleteGeoMatchSetResponse_changeToken - The ChangeToken that you used to submit the DeleteGeoMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:DeleteGeoMatchSetResponse', deleteGeoMatchSetResponse_httpStatus - The response's http status code.

Response Lenses

deleteGeoMatchSetResponse_changeToken :: Lens' DeleteGeoMatchSetResponse (Maybe Text)

The ChangeToken that you used to submit the DeleteGeoMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

deleteGeoMatchSetResponse_httpStatus :: Lens' DeleteGeoMatchSetResponse Int

The response's http status code.