Network.AWS.WAFRegional.DeleteRuleGroup

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Permanently deletes a RuleGroup. You can't delete a RuleGroup if it's still used in any WebACL objects or if it still includes any rules.

If you just want to remove a RuleGroup from a WebACL, use UpdateWebACL.

To permanently delete a RuleGroup from AWS WAF, perform the following steps:

Update the RuleGroup to remove rules, if any. For more information, see UpdateRuleGroup.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a DeleteRuleGroup request.

Submit a DeleteRuleGroup request.

Creating a Request

data DeleteRuleGroup

See: newDeleteRuleGroup smart constructor.

Constructors

DeleteRuleGroup'

Fields

ruleGroupId :: Text

The RuleGroupId of the RuleGroup that you want to delete. RuleGroupId is returned by CreateRuleGroup and by ListRuleGroups.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

newDeleteRuleGroup

Arguments

:: Text
$sel:ruleGroupId:DeleteRuleGroup'

-> Text
$sel:changeToken:DeleteRuleGroup'

-> DeleteRuleGroup

Create a value of DeleteRuleGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleGroupId:DeleteRuleGroup', deleteRuleGroup_ruleGroupId - The RuleGroupId of the RuleGroup that you want to delete. RuleGroupId is returned by CreateRuleGroup and by ListRuleGroups.

$sel:changeToken:DeleteRuleGroup', deleteRuleGroup_changeToken - The value returned by the most recent call to GetChangeToken.

Request Lenses

deleteRuleGroup_ruleGroupId :: Lens' DeleteRuleGroup Text

The RuleGroupId of the RuleGroup that you want to delete. RuleGroupId is returned by CreateRuleGroup and by ListRuleGroups.

deleteRuleGroup_changeToken :: Lens' DeleteRuleGroup Text

The value returned by the most recent call to GetChangeToken.

Destructuring the Response

data DeleteRuleGroupResponse

See: newDeleteRuleGroupResponse smart constructor.

Constructors

DeleteRuleGroupResponse'

Fields

changeToken :: Maybe Text

The ChangeToken that you used to submit the DeleteRuleGroup request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newDeleteRuleGroupResponse

Arguments

:: Int
$sel:httpStatus:DeleteRuleGroupResponse'

-> DeleteRuleGroupResponse

Create a value of DeleteRuleGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeToken:DeleteRuleGroup', deleteRuleGroupResponse_changeToken - The ChangeToken that you used to submit the DeleteRuleGroup request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:DeleteRuleGroupResponse', deleteRuleGroupResponse_httpStatus - The response's http status code.

Response Lenses

deleteRuleGroupResponse_changeToken :: Lens' DeleteRuleGroupResponse (Maybe Text)

The ChangeToken that you used to submit the DeleteRuleGroup request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

deleteRuleGroupResponse_httpStatus :: Lens' DeleteRuleGroupResponse Int

The response's http status code.