Network.AWS.WAFRegional.DeleteSqlInjectionMatchSet

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Permanently deletes a SqlInjectionMatchSet. You can't delete a SqlInjectionMatchSet if it's still used in any Rules or if it still contains any SqlInjectionMatchTuple objects.

If you just want to remove a SqlInjectionMatchSet from a Rule, use UpdateRule.

To permanently delete a SqlInjectionMatchSet from AWS WAF, perform the following steps:

Update the SqlInjectionMatchSet to remove filters, if any. For more information, see UpdateSqlInjectionMatchSet.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a DeleteSqlInjectionMatchSet request.

Submit a DeleteSqlInjectionMatchSet request.

Creating a Request

data DeleteSqlInjectionMatchSet

A request to delete a SqlInjectionMatchSet from AWS WAF.

See: newDeleteSqlInjectionMatchSet smart constructor.

Constructors

DeleteSqlInjectionMatchSet'

Fields

sqlInjectionMatchSetId :: Text

The SqlInjectionMatchSetId of the SqlInjectionMatchSet that you want to delete. SqlInjectionMatchSetId is returned by CreateSqlInjectionMatchSet and by ListSqlInjectionMatchSets.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

newDeleteSqlInjectionMatchSet

Arguments

:: Text
$sel:sqlInjectionMatchSetId:DeleteSqlInjectionMatchSet'

-> Text
$sel:changeToken:DeleteSqlInjectionMatchSet'

-> DeleteSqlInjectionMatchSet

Create a value of DeleteSqlInjectionMatchSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sqlInjectionMatchSetId:DeleteSqlInjectionMatchSet', deleteSqlInjectionMatchSet_sqlInjectionMatchSetId - The SqlInjectionMatchSetId of the SqlInjectionMatchSet that you want to delete. SqlInjectionMatchSetId is returned by CreateSqlInjectionMatchSet and by ListSqlInjectionMatchSets.

$sel:changeToken:DeleteSqlInjectionMatchSet', deleteSqlInjectionMatchSet_changeToken - The value returned by the most recent call to GetChangeToken.

Request Lenses

deleteSqlInjectionMatchSet_sqlInjectionMatchSetId :: Lens' DeleteSqlInjectionMatchSet Text

The SqlInjectionMatchSetId of the SqlInjectionMatchSet that you want to delete. SqlInjectionMatchSetId is returned by CreateSqlInjectionMatchSet and by ListSqlInjectionMatchSets.

deleteSqlInjectionMatchSet_changeToken :: Lens' DeleteSqlInjectionMatchSet Text

The value returned by the most recent call to GetChangeToken.

Destructuring the Response

data DeleteSqlInjectionMatchSetResponse

The response to a request to delete a SqlInjectionMatchSet from AWS WAF.

See: newDeleteSqlInjectionMatchSetResponse smart constructor.

Constructors

DeleteSqlInjectionMatchSetResponse'

Fields

changeToken :: Maybe Text

The ChangeToken that you used to submit the DeleteSqlInjectionMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newDeleteSqlInjectionMatchSetResponse

Arguments

:: Int
$sel:httpStatus:DeleteSqlInjectionMatchSetResponse'

-> DeleteSqlInjectionMatchSetResponse

Create a value of DeleteSqlInjectionMatchSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeToken:DeleteSqlInjectionMatchSet', deleteSqlInjectionMatchSetResponse_changeToken - The ChangeToken that you used to submit the DeleteSqlInjectionMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:DeleteSqlInjectionMatchSetResponse', deleteSqlInjectionMatchSetResponse_httpStatus - The response's http status code.

Response Lenses

deleteSqlInjectionMatchSetResponse_changeToken :: Lens' DeleteSqlInjectionMatchSetResponse (Maybe Text)

The ChangeToken that you used to submit the DeleteSqlInjectionMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

deleteSqlInjectionMatchSetResponse_httpStatus :: Lens' DeleteSqlInjectionMatchSetResponse Int

The response's http status code.