Network.AWS.WAFRegional.GetRateBasedRule

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Returns the RateBasedRule that is specified by the RuleId that you included in the GetRateBasedRule request.

Creating a Request

data GetRateBasedRule

See: newGetRateBasedRule smart constructor.

Constructors

GetRateBasedRule'

Fields

ruleId :: Text

The RuleId of the RateBasedRule that you want to get. RuleId is returned by CreateRateBasedRule and by ListRateBasedRules.

newGetRateBasedRule

Arguments

:: Text
$sel:ruleId:GetRateBasedRule'

-> GetRateBasedRule

Create a value of GetRateBasedRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleId:GetRateBasedRule', getRateBasedRule_ruleId - The RuleId of the RateBasedRule that you want to get. RuleId is returned by CreateRateBasedRule and by ListRateBasedRules.

Request Lenses

getRateBasedRule_ruleId :: Lens' GetRateBasedRule Text

The RuleId of the RateBasedRule that you want to get. RuleId is returned by CreateRateBasedRule and by ListRateBasedRules.

Destructuring the Response

data GetRateBasedRuleResponse

See: newGetRateBasedRuleResponse smart constructor.

Constructors

GetRateBasedRuleResponse'

Fields

rule :: Maybe RateBasedRule

Information about the RateBasedRule that you specified in the GetRateBasedRule request.

httpStatus :: Int

The response's http status code.

newGetRateBasedRuleResponse

Arguments

:: Int
$sel:httpStatus:GetRateBasedRuleResponse'

-> GetRateBasedRuleResponse

Create a value of GetRateBasedRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rule:GetRateBasedRuleResponse', getRateBasedRuleResponse_rule - Information about the RateBasedRule that you specified in the GetRateBasedRule request.

$sel:httpStatus:GetRateBasedRuleResponse', getRateBasedRuleResponse_httpStatus - The response's http status code.

Response Lenses

getRateBasedRuleResponse_rule :: Lens' GetRateBasedRuleResponse (Maybe RateBasedRule)

Information about the RateBasedRule that you specified in the GetRateBasedRule request.

getRateBasedRuleResponse_httpStatus :: Lens' GetRateBasedRuleResponse Int

The response's http status code.