Network.AWS.WAFRegional.GetRegexPatternSet

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Returns the RegexPatternSet specified by RegexPatternSetId.

Creating a Request

data GetRegexPatternSet

See: newGetRegexPatternSet smart constructor.

Constructors

GetRegexPatternSet'

Fields

regexPatternSetId :: Text

The RegexPatternSetId of the RegexPatternSet that you want to get. RegexPatternSetId is returned by CreateRegexPatternSet and by ListRegexPatternSets.

newGetRegexPatternSet

Arguments

:: Text
$sel:regexPatternSetId:GetRegexPatternSet'

-> GetRegexPatternSet

Create a value of GetRegexPatternSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:regexPatternSetId:GetRegexPatternSet', getRegexPatternSet_regexPatternSetId - The RegexPatternSetId of the RegexPatternSet that you want to get. RegexPatternSetId is returned by CreateRegexPatternSet and by ListRegexPatternSets.

Request Lenses

getRegexPatternSet_regexPatternSetId :: Lens' GetRegexPatternSet Text

The RegexPatternSetId of the RegexPatternSet that you want to get. RegexPatternSetId is returned by CreateRegexPatternSet and by ListRegexPatternSets.

Destructuring the Response

data GetRegexPatternSetResponse

See: newGetRegexPatternSetResponse smart constructor.

Constructors

GetRegexPatternSetResponse'

Fields

regexPatternSet :: Maybe RegexPatternSet

Information about the RegexPatternSet that you specified in the GetRegexPatternSet request, including the identifier of the pattern set and the regular expression patterns you want AWS WAF to search for.

httpStatus :: Int

The response's http status code.

newGetRegexPatternSetResponse

Arguments

:: Int
$sel:httpStatus:GetRegexPatternSetResponse'

-> GetRegexPatternSetResponse

Create a value of GetRegexPatternSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:regexPatternSet:GetRegexPatternSetResponse', getRegexPatternSetResponse_regexPatternSet - Information about the RegexPatternSet that you specified in the GetRegexPatternSet request, including the identifier of the pattern set and the regular expression patterns you want AWS WAF to search for.

$sel:httpStatus:GetRegexPatternSetResponse', getRegexPatternSetResponse_httpStatus - The response's http status code.

Response Lenses

getRegexPatternSetResponse_regexPatternSet :: Lens' GetRegexPatternSetResponse (Maybe RegexPatternSet)

Information about the RegexPatternSet that you specified in the GetRegexPatternSet request, including the identifier of the pattern set and the regular expression patterns you want AWS WAF to search for.

getRegexPatternSetResponse_httpStatus :: Lens' GetRegexPatternSetResponse Int

The response's http status code.