Network.AWS.WAFRegional.ListLoggingConfigurations

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Returns an array of LoggingConfiguration objects.

Creating a Request

data ListLoggingConfigurations

See: newListLoggingConfigurations smart constructor.

Constructors

ListLoggingConfigurations'

Fields

nextMarker :: Maybe Text

If you specify a value for Limit and you have more LoggingConfigurations than the value of Limit, AWS WAF returns a NextMarker value in the response that allows you to list another group of LoggingConfigurations. For the second and subsequent ListLoggingConfigurations requests, specify the value of NextMarker from the previous response to get information about another batch of ListLoggingConfigurations.

limit :: Maybe Natural

Specifies the number of LoggingConfigurations that you want AWS WAF to return for this request. If you have more LoggingConfigurations than the number that you specify for Limit, the response includes a NextMarker value that you can use to get another batch of LoggingConfigurations.

newListLoggingConfigurations :: ListLoggingConfigurations

Create a value of ListLoggingConfigurations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextMarker:ListLoggingConfigurations', listLoggingConfigurations_nextMarker - If you specify a value for Limit and you have more LoggingConfigurations than the value of Limit, AWS WAF returns a NextMarker value in the response that allows you to list another group of LoggingConfigurations. For the second and subsequent ListLoggingConfigurations requests, specify the value of NextMarker from the previous response to get information about another batch of ListLoggingConfigurations.

$sel:limit:ListLoggingConfigurations', listLoggingConfigurations_limit - Specifies the number of LoggingConfigurations that you want AWS WAF to return for this request. If you have more LoggingConfigurations than the number that you specify for Limit, the response includes a NextMarker value that you can use to get another batch of LoggingConfigurations.

Request Lenses

listLoggingConfigurations_nextMarker :: Lens' ListLoggingConfigurations (Maybe Text)

If you specify a value for Limit and you have more LoggingConfigurations than the value of Limit, AWS WAF returns a NextMarker value in the response that allows you to list another group of LoggingConfigurations. For the second and subsequent ListLoggingConfigurations requests, specify the value of NextMarker from the previous response to get information about another batch of ListLoggingConfigurations.

listLoggingConfigurations_limit :: Lens' ListLoggingConfigurations (Maybe Natural)

Specifies the number of LoggingConfigurations that you want AWS WAF to return for this request. If you have more LoggingConfigurations than the number that you specify for Limit, the response includes a NextMarker value that you can use to get another batch of LoggingConfigurations.

Destructuring the Response

data ListLoggingConfigurationsResponse

See: newListLoggingConfigurationsResponse smart constructor.

Constructors

ListLoggingConfigurationsResponse'

Fields

nextMarker :: Maybe Text

If you have more LoggingConfigurations than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more LoggingConfigurations, submit another ListLoggingConfigurations request, and specify the NextMarker value from the response in the NextMarker value in the next request.

loggingConfigurations :: Maybe [LoggingConfiguration]

An array of LoggingConfiguration objects.

httpStatus :: Int

The response's http status code.

newListLoggingConfigurationsResponse

Arguments

:: Int
$sel:httpStatus:ListLoggingConfigurationsResponse'

-> ListLoggingConfigurationsResponse

Create a value of ListLoggingConfigurationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextMarker:ListLoggingConfigurations', listLoggingConfigurationsResponse_nextMarker - If you have more LoggingConfigurations than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more LoggingConfigurations, submit another ListLoggingConfigurations request, and specify the NextMarker value from the response in the NextMarker value in the next request.

$sel:loggingConfigurations:ListLoggingConfigurationsResponse', listLoggingConfigurationsResponse_loggingConfigurations - An array of LoggingConfiguration objects.

$sel:httpStatus:ListLoggingConfigurationsResponse', listLoggingConfigurationsResponse_httpStatus - The response's http status code.

Response Lenses

listLoggingConfigurationsResponse_nextMarker :: Lens' ListLoggingConfigurationsResponse (Maybe Text)

If you have more LoggingConfigurations than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more LoggingConfigurations, submit another ListLoggingConfigurations request, and specify the NextMarker value from the response in the NextMarker value in the next request.

listLoggingConfigurationsResponse_loggingConfigurations :: Lens' ListLoggingConfigurationsResponse (Maybe [LoggingConfiguration])

An array of LoggingConfiguration objects.

listLoggingConfigurationsResponse_httpStatus :: Lens' ListLoggingConfigurationsResponse Int

The response's http status code.