Network.AWS.WAFRegional.ListRegexMatchSets

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Returns an array of RegexMatchSetSummary objects.

Creating a Request

data ListRegexMatchSets

See: newListRegexMatchSets smart constructor.

Constructors

ListRegexMatchSets'

Fields

nextMarker :: Maybe Text

If you specify a value for Limit and you have more RegexMatchSet objects than the value of Limit, AWS WAF returns a NextMarker value in the response that allows you to list another group of ByteMatchSets. For the second and subsequent ListRegexMatchSets requests, specify the value of NextMarker from the previous response to get information about another batch of RegexMatchSet objects.

limit :: Maybe Natural

Specifies the number of RegexMatchSet objects that you want AWS WAF to return for this request. If you have more RegexMatchSet objects than the number you specify for Limit, the response includes a NextMarker value that you can use to get another batch of RegexMatchSet objects.

newListRegexMatchSets :: ListRegexMatchSets

Create a value of ListRegexMatchSets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextMarker:ListRegexMatchSets', listRegexMatchSets_nextMarker - If you specify a value for Limit and you have more RegexMatchSet objects than the value of Limit, AWS WAF returns a NextMarker value in the response that allows you to list another group of ByteMatchSets. For the second and subsequent ListRegexMatchSets requests, specify the value of NextMarker from the previous response to get information about another batch of RegexMatchSet objects.

$sel:limit:ListRegexMatchSets', listRegexMatchSets_limit - Specifies the number of RegexMatchSet objects that you want AWS WAF to return for this request. If you have more RegexMatchSet objects than the number you specify for Limit, the response includes a NextMarker value that you can use to get another batch of RegexMatchSet objects.

Request Lenses

listRegexMatchSets_nextMarker :: Lens' ListRegexMatchSets (Maybe Text)

If you specify a value for Limit and you have more RegexMatchSet objects than the value of Limit, AWS WAF returns a NextMarker value in the response that allows you to list another group of ByteMatchSets. For the second and subsequent ListRegexMatchSets requests, specify the value of NextMarker from the previous response to get information about another batch of RegexMatchSet objects.

listRegexMatchSets_limit :: Lens' ListRegexMatchSets (Maybe Natural)

Specifies the number of RegexMatchSet objects that you want AWS WAF to return for this request. If you have more RegexMatchSet objects than the number you specify for Limit, the response includes a NextMarker value that you can use to get another batch of RegexMatchSet objects.

Destructuring the Response

data ListRegexMatchSetsResponse

See: newListRegexMatchSetsResponse smart constructor.

Constructors

ListRegexMatchSetsResponse'

Fields

nextMarker :: Maybe Text

If you have more RegexMatchSet objects than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more RegexMatchSet objects, submit another ListRegexMatchSets request, and specify the NextMarker value from the response in the NextMarker value in the next request.

regexMatchSets :: Maybe [RegexMatchSetSummary]

An array of RegexMatchSetSummary objects.

httpStatus :: Int

The response's http status code.

newListRegexMatchSetsResponse

Arguments

:: Int
$sel:httpStatus:ListRegexMatchSetsResponse'

-> ListRegexMatchSetsResponse

Create a value of ListRegexMatchSetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextMarker:ListRegexMatchSets', listRegexMatchSetsResponse_nextMarker - If you have more RegexMatchSet objects than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more RegexMatchSet objects, submit another ListRegexMatchSets request, and specify the NextMarker value from the response in the NextMarker value in the next request.

$sel:regexMatchSets:ListRegexMatchSetsResponse', listRegexMatchSetsResponse_regexMatchSets - An array of RegexMatchSetSummary objects.

$sel:httpStatus:ListRegexMatchSetsResponse', listRegexMatchSetsResponse_httpStatus - The response's http status code.

Response Lenses

listRegexMatchSetsResponse_nextMarker :: Lens' ListRegexMatchSetsResponse (Maybe Text)

If you have more RegexMatchSet objects than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more RegexMatchSet objects, submit another ListRegexMatchSets request, and specify the NextMarker value from the response in the NextMarker value in the next request.

listRegexMatchSetsResponse_regexMatchSets :: Lens' ListRegexMatchSetsResponse (Maybe [RegexMatchSetSummary])

An array of RegexMatchSetSummary objects.

listRegexMatchSetsResponse_httpStatus :: Lens' ListRegexMatchSetsResponse Int

The response's http status code.