Network.AWS.WAFRegional.Types.ByteMatchSet

Description

Documentation

data ByteMatchSet

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

In a GetByteMatchSet request, ByteMatchSet is a complex type that contains the ByteMatchSetId and Name of a ByteMatchSet, and the values that you specified when you updated the ByteMatchSet.

A complex type that contains ByteMatchTuple objects, which specify the parts of web requests that you want AWS WAF to inspect and the values that you want AWS WAF to search for. If a ByteMatchSet contains more than one ByteMatchTuple object, a request needs to match the settings in only one ByteMatchTuple to be considered a match.

See: newByteMatchSet smart constructor.

Constructors

ByteMatchSet'

Fields

name :: Maybe Text

A friendly name or description of the ByteMatchSet. You can't change Name after you create a ByteMatchSet.

byteMatchSetId :: Text

The ByteMatchSetId for a ByteMatchSet. You use ByteMatchSetId to get information about a ByteMatchSet (see GetByteMatchSet), update a ByteMatchSet (see UpdateByteMatchSet), insert a ByteMatchSet into a Rule or delete one from a Rule (see UpdateRule), and delete a ByteMatchSet from AWS WAF (see DeleteByteMatchSet).

ByteMatchSetId is returned by CreateByteMatchSet and by ListByteMatchSets.

byteMatchTuples :: [ByteMatchTuple]

Specifies the bytes (typically a string that corresponds with ASCII characters) that you want AWS WAF to search for in web requests, the location in requests that you want AWS WAF to search, and other settings.

newByteMatchSet

Arguments

:: Text
$sel:byteMatchSetId:ByteMatchSet'

-> ByteMatchSet

Create a value of ByteMatchSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ByteMatchSet', byteMatchSet_name - A friendly name or description of the ByteMatchSet. You can't change Name after you create a ByteMatchSet.

$sel:byteMatchSetId:ByteMatchSet', byteMatchSet_byteMatchSetId - The ByteMatchSetId for a ByteMatchSet. You use ByteMatchSetId to get information about a ByteMatchSet (see GetByteMatchSet), update a ByteMatchSet (see UpdateByteMatchSet), insert a ByteMatchSet into a Rule or delete one from a Rule (see UpdateRule), and delete a ByteMatchSet from AWS WAF (see DeleteByteMatchSet).

ByteMatchSetId is returned by CreateByteMatchSet and by ListByteMatchSets.

$sel:byteMatchTuples:ByteMatchSet', byteMatchSet_byteMatchTuples - Specifies the bytes (typically a string that corresponds with ASCII characters) that you want AWS WAF to search for in web requests, the location in requests that you want AWS WAF to search, and other settings.

byteMatchSet_name :: Lens' ByteMatchSet (Maybe Text)

A friendly name or description of the ByteMatchSet. You can't change Name after you create a ByteMatchSet.

byteMatchSet_byteMatchSetId :: Lens' ByteMatchSet Text

The ByteMatchSetId for a ByteMatchSet. You use ByteMatchSetId to get information about a ByteMatchSet (see GetByteMatchSet), update a ByteMatchSet (see UpdateByteMatchSet), insert a ByteMatchSet into a Rule or delete one from a Rule (see UpdateRule), and delete a ByteMatchSet from AWS WAF (see DeleteByteMatchSet).

ByteMatchSetId is returned by CreateByteMatchSet and by ListByteMatchSets.

byteMatchSet_byteMatchTuples :: Lens' ByteMatchSet [ByteMatchTuple]

Specifies the bytes (typically a string that corresponds with ASCII characters) that you want AWS WAF to search for in web requests, the location in requests that you want AWS WAF to search, and other settings.