Network.AWS.WAFRegional.Types.ByteMatchSetSummary

Description

Documentation

data ByteMatchSetSummary

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Returned by ListByteMatchSets. Each ByteMatchSetSummary object includes the Name and ByteMatchSetId for one ByteMatchSet.

See: newByteMatchSetSummary smart constructor.

Constructors

ByteMatchSetSummary'

Fields

byteMatchSetId :: Text

The ByteMatchSetId for a ByteMatchSet. You use ByteMatchSetId to get information about a ByteMatchSet, update a ByteMatchSet, remove a ByteMatchSet from a Rule, and delete a ByteMatchSet from AWS WAF.

ByteMatchSetId is returned by CreateByteMatchSet and by ListByteMatchSets.

name :: Text

A friendly name or description of the ByteMatchSet. You can't change Name after you create a ByteMatchSet.

newByteMatchSetSummary

Arguments

:: Text
$sel:byteMatchSetId:ByteMatchSetSummary'

-> Text
$sel:name:ByteMatchSetSummary'

-> ByteMatchSetSummary

Create a value of ByteMatchSetSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:byteMatchSetId:ByteMatchSetSummary', byteMatchSetSummary_byteMatchSetId - The ByteMatchSetId for a ByteMatchSet. You use ByteMatchSetId to get information about a ByteMatchSet, update a ByteMatchSet, remove a ByteMatchSet from a Rule, and delete a ByteMatchSet from AWS WAF.

ByteMatchSetId is returned by CreateByteMatchSet and by ListByteMatchSets.

$sel:name:ByteMatchSetSummary', byteMatchSetSummary_name - A friendly name or description of the ByteMatchSet. You can't change Name after you create a ByteMatchSet.

byteMatchSetSummary_byteMatchSetId :: Lens' ByteMatchSetSummary Text

The ByteMatchSetId for a ByteMatchSet. You use ByteMatchSetId to get information about a ByteMatchSet, update a ByteMatchSet, remove a ByteMatchSet from a Rule, and delete a ByteMatchSet from AWS WAF.

ByteMatchSetId is returned by CreateByteMatchSet and by ListByteMatchSets.

byteMatchSetSummary_name :: Lens' ByteMatchSetSummary Text

A friendly name or description of the ByteMatchSet. You can't change Name after you create a ByteMatchSet.