Network.AWS.WAFRegional.Types.SampledHTTPRequest

Description

Documentation

data SampledHTTPRequest

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

The response from a GetSampledRequests request includes a SampledHTTPRequests complex type that appears as SampledRequests in the response syntax. SampledHTTPRequests contains one SampledHTTPRequest object for each web request that is returned by GetSampledRequests.

See: newSampledHTTPRequest smart constructor.

Constructors

SampledHTTPRequest'

Fields

action :: Maybe Text

The action for the Rule that the request matched: ALLOW, BLOCK, or COUNT.

timestamp :: Maybe POSIX

The time at which AWS WAF received the request from your AWS resource, in Unix time format (in seconds).

ruleWithinRuleGroup :: Maybe Text

This value is returned if the GetSampledRequests request specifies the ID of a RuleGroup rather than the ID of an individual rule. RuleWithinRuleGroup is the rule within the specified RuleGroup that matched the request listed in the response.

request :: HTTPRequest

A complex type that contains detailed information about the request.

weight :: Natural

A value that indicates how one result in the response relates proportionally to other results in the response. A result that has a weight of 2 represents roughly twice as many CloudFront web requests as a result that has a weight of 1.

newSampledHTTPRequest

Arguments

:: HTTPRequest
$sel:request:SampledHTTPRequest'

-> Natural
$sel:weight:SampledHTTPRequest'

-> SampledHTTPRequest

Create a value of SampledHTTPRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:action:SampledHTTPRequest', sampledHTTPRequest_action - The action for the Rule that the request matched: ALLOW, BLOCK, or COUNT.

$sel:timestamp:SampledHTTPRequest', sampledHTTPRequest_timestamp - The time at which AWS WAF received the request from your AWS resource, in Unix time format (in seconds).

$sel:ruleWithinRuleGroup:SampledHTTPRequest', sampledHTTPRequest_ruleWithinRuleGroup - This value is returned if the GetSampledRequests request specifies the ID of a RuleGroup rather than the ID of an individual rule. RuleWithinRuleGroup is the rule within the specified RuleGroup that matched the request listed in the response.

$sel:request:SampledHTTPRequest', sampledHTTPRequest_request - A complex type that contains detailed information about the request.

$sel:weight:SampledHTTPRequest', sampledHTTPRequest_weight - A value that indicates how one result in the response relates proportionally to other results in the response. A result that has a weight of 2 represents roughly twice as many CloudFront web requests as a result that has a weight of 1.

sampledHTTPRequest_action :: Lens' SampledHTTPRequest (Maybe Text)

The action for the Rule that the request matched: ALLOW, BLOCK, or COUNT.

sampledHTTPRequest_timestamp :: Lens' SampledHTTPRequest (Maybe UTCTime)

The time at which AWS WAF received the request from your AWS resource, in Unix time format (in seconds).

sampledHTTPRequest_ruleWithinRuleGroup :: Lens' SampledHTTPRequest (Maybe Text)

This value is returned if the GetSampledRequests request specifies the ID of a RuleGroup rather than the ID of an individual rule. RuleWithinRuleGroup is the rule within the specified RuleGroup that matched the request listed in the response.

sampledHTTPRequest_request :: Lens' SampledHTTPRequest HTTPRequest

A complex type that contains detailed information about the request.

sampledHTTPRequest_weight :: Lens' SampledHTTPRequest Natural

A value that indicates how one result in the response relates proportionally to other results in the response. A result that has a weight of 2 represents roughly twice as many CloudFront web requests as a result that has a weight of 1.