Network.AWS.WAFRegional.Types.SizeConstraintSet

Description

Documentation

data SizeConstraintSet

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

A complex type that contains SizeConstraint objects, which specify the parts of web requests that you want AWS WAF to inspect the size of. If a SizeConstraintSet contains more than one SizeConstraint object, a request only needs to match one constraint to be considered a match.

See: newSizeConstraintSet smart constructor.

Constructors

SizeConstraintSet'

Fields

name :: Maybe Text

The name, if any, of the SizeConstraintSet.

sizeConstraintSetId :: Text

A unique identifier for a SizeConstraintSet. You use SizeConstraintSetId to get information about a SizeConstraintSet (see GetSizeConstraintSet), update a SizeConstraintSet (see UpdateSizeConstraintSet), insert a SizeConstraintSet into a Rule or delete one from a Rule (see UpdateRule), and delete a SizeConstraintSet from AWS WAF (see DeleteSizeConstraintSet).

SizeConstraintSetId is returned by CreateSizeConstraintSet and by ListSizeConstraintSets.

sizeConstraints :: [SizeConstraint]

Specifies the parts of web requests that you want to inspect the size of.

newSizeConstraintSet

Arguments

:: Text
$sel:sizeConstraintSetId:SizeConstraintSet'

-> SizeConstraintSet

Create a value of SizeConstraintSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:SizeConstraintSet', sizeConstraintSet_name - The name, if any, of the SizeConstraintSet.

$sel:sizeConstraintSetId:SizeConstraintSet', sizeConstraintSet_sizeConstraintSetId - A unique identifier for a SizeConstraintSet. You use SizeConstraintSetId to get information about a SizeConstraintSet (see GetSizeConstraintSet), update a SizeConstraintSet (see UpdateSizeConstraintSet), insert a SizeConstraintSet into a Rule or delete one from a Rule (see UpdateRule), and delete a SizeConstraintSet from AWS WAF (see DeleteSizeConstraintSet).

SizeConstraintSetId is returned by CreateSizeConstraintSet and by ListSizeConstraintSets.

$sel:sizeConstraints:SizeConstraintSet', sizeConstraintSet_sizeConstraints - Specifies the parts of web requests that you want to inspect the size of.

sizeConstraintSet_name :: Lens' SizeConstraintSet (Maybe Text)

The name, if any, of the SizeConstraintSet.

sizeConstraintSet_sizeConstraintSetId :: Lens' SizeConstraintSet Text

A unique identifier for a SizeConstraintSet. You use SizeConstraintSetId to get information about a SizeConstraintSet (see GetSizeConstraintSet), update a SizeConstraintSet (see UpdateSizeConstraintSet), insert a SizeConstraintSet into a Rule or delete one from a Rule (see UpdateRule), and delete a SizeConstraintSet from AWS WAF (see DeleteSizeConstraintSet).

SizeConstraintSetId is returned by CreateSizeConstraintSet and by ListSizeConstraintSets.

sizeConstraintSet_sizeConstraints :: Lens' SizeConstraintSet [SizeConstraint]

Specifies the parts of web requests that you want to inspect the size of.