Network.AWS.WAFRegional.Types.SizeConstraintSetSummary

Description

Documentation

data SizeConstraintSetSummary

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

The Id and Name of a SizeConstraintSet.

See: newSizeConstraintSetSummary smart constructor.

Constructors

SizeConstraintSetSummary'

Fields

sizeConstraintSetId :: Text

A unique identifier for a SizeConstraintSet. You use SizeConstraintSetId to get information about a SizeConstraintSet (see GetSizeConstraintSet), update a SizeConstraintSet (see UpdateSizeConstraintSet), insert a SizeConstraintSet into a Rule or delete one from a Rule (see UpdateRule), and delete a SizeConstraintSet from AWS WAF (see DeleteSizeConstraintSet).

SizeConstraintSetId is returned by CreateSizeConstraintSet and by ListSizeConstraintSets.

name :: Text

The name of the SizeConstraintSet, if any.

newSizeConstraintSetSummary

Arguments

:: Text
$sel:sizeConstraintSetId:SizeConstraintSetSummary'

-> Text
$sel:name:SizeConstraintSetSummary'

-> SizeConstraintSetSummary

Create a value of SizeConstraintSetSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sizeConstraintSetId:SizeConstraintSetSummary', sizeConstraintSetSummary_sizeConstraintSetId - A unique identifier for a SizeConstraintSet. You use SizeConstraintSetId to get information about a SizeConstraintSet (see GetSizeConstraintSet), update a SizeConstraintSet (see UpdateSizeConstraintSet), insert a SizeConstraintSet into a Rule or delete one from a Rule (see UpdateRule), and delete a SizeConstraintSet from AWS WAF (see DeleteSizeConstraintSet).

SizeConstraintSetId is returned by CreateSizeConstraintSet and by ListSizeConstraintSets.

$sel:name:SizeConstraintSetSummary', sizeConstraintSetSummary_name - The name of the SizeConstraintSet, if any.

sizeConstraintSetSummary_sizeConstraintSetId :: Lens' SizeConstraintSetSummary Text

A unique identifier for a SizeConstraintSet. You use SizeConstraintSetId to get information about a SizeConstraintSet (see GetSizeConstraintSet), update a SizeConstraintSet (see UpdateSizeConstraintSet), insert a SizeConstraintSet into a Rule or delete one from a Rule (see UpdateRule), and delete a SizeConstraintSet from AWS WAF (see DeleteSizeConstraintSet).

SizeConstraintSetId is returned by CreateSizeConstraintSet and by ListSizeConstraintSets.

sizeConstraintSetSummary_name :: Lens' SizeConstraintSetSummary Text

The name of the SizeConstraintSet, if any.