Network.AWS.WAFRegional.Types.SizeConstraintSetUpdate

Description

Documentation

data SizeConstraintSetUpdate

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Specifies the part of a web request that you want to inspect the size of and indicates whether you want to add the specification to a SizeConstraintSet or delete it from a SizeConstraintSet.

See: newSizeConstraintSetUpdate smart constructor.

Constructors

SizeConstraintSetUpdate'

Fields

action :: ChangeAction

Specify INSERT to add a SizeConstraintSetUpdate to a SizeConstraintSet. Use DELETE to remove a SizeConstraintSetUpdate from a SizeConstraintSet.

sizeConstraint :: SizeConstraint

Specifies a constraint on the size of a part of the web request. AWS WAF uses the Size, ComparisonOperator, and FieldToMatch to build an expression in the form of "Size ComparisonOperator size in bytes of FieldToMatch". If that expression is true, the SizeConstraint is considered to match.

newSizeConstraintSetUpdate

Arguments

:: ChangeAction
$sel:action:SizeConstraintSetUpdate'

-> SizeConstraint
$sel:sizeConstraint:SizeConstraintSetUpdate'

-> SizeConstraintSetUpdate

Create a value of SizeConstraintSetUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:action:SizeConstraintSetUpdate', sizeConstraintSetUpdate_action - Specify INSERT to add a SizeConstraintSetUpdate to a SizeConstraintSet. Use DELETE to remove a SizeConstraintSetUpdate from a SizeConstraintSet.

$sel:sizeConstraint:SizeConstraintSetUpdate', sizeConstraintSetUpdate_sizeConstraint - Specifies a constraint on the size of a part of the web request. AWS WAF uses the Size, ComparisonOperator, and FieldToMatch to build an expression in the form of "Size ComparisonOperator size in bytes of FieldToMatch". If that expression is true, the SizeConstraint is considered to match.

sizeConstraintSetUpdate_action :: Lens' SizeConstraintSetUpdate ChangeAction

Specify INSERT to add a SizeConstraintSetUpdate to a SizeConstraintSet. Use DELETE to remove a SizeConstraintSetUpdate from a SizeConstraintSet.

sizeConstraintSetUpdate_sizeConstraint :: Lens' SizeConstraintSetUpdate SizeConstraint

Specifies a constraint on the size of a part of the web request. AWS WAF uses the Size, ComparisonOperator, and FieldToMatch to build an expression in the form of "Size ComparisonOperator size in bytes of FieldToMatch". If that expression is true, the SizeConstraint is considered to match.