Network.AWS.WAFRegional.Types.XssMatchSet

Description

Documentation

data XssMatchSet

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

A complex type that contains XssMatchTuple objects, which specify the parts of web requests that you want AWS WAF to inspect for cross-site scripting attacks and, if you want AWS WAF to inspect a header, the name of the header. If a XssMatchSet contains more than one XssMatchTuple object, a request needs to include cross-site scripting attacks in only one of the specified parts of the request to be considered a match.

See: newXssMatchSet smart constructor.

Constructors

XssMatchSet'

Fields

name :: Maybe Text

The name, if any, of the XssMatchSet.

xssMatchSetId :: Text

A unique identifier for an XssMatchSet. You use XssMatchSetId to get information about an XssMatchSet (see GetXssMatchSet), update an XssMatchSet (see UpdateXssMatchSet), insert an XssMatchSet into a Rule or delete one from a Rule (see UpdateRule), and delete an XssMatchSet from AWS WAF (see DeleteXssMatchSet).

XssMatchSetId is returned by CreateXssMatchSet and by ListXssMatchSets.

xssMatchTuples :: [XssMatchTuple]

Specifies the parts of web requests that you want to inspect for cross-site scripting attacks.

newXssMatchSet

Arguments

:: Text
$sel:xssMatchSetId:XssMatchSet'

-> XssMatchSet

Create a value of XssMatchSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:XssMatchSet', xssMatchSet_name - The name, if any, of the XssMatchSet.

$sel:xssMatchSetId:XssMatchSet', xssMatchSet_xssMatchSetId - A unique identifier for an XssMatchSet. You use XssMatchSetId to get information about an XssMatchSet (see GetXssMatchSet), update an XssMatchSet (see UpdateXssMatchSet), insert an XssMatchSet into a Rule or delete one from a Rule (see UpdateRule), and delete an XssMatchSet from AWS WAF (see DeleteXssMatchSet).

XssMatchSetId is returned by CreateXssMatchSet and by ListXssMatchSets.

$sel:xssMatchTuples:XssMatchSet', xssMatchSet_xssMatchTuples - Specifies the parts of web requests that you want to inspect for cross-site scripting attacks.

xssMatchSet_name :: Lens' XssMatchSet (Maybe Text)

The name, if any, of the XssMatchSet.

xssMatchSet_xssMatchSetId :: Lens' XssMatchSet Text

A unique identifier for an XssMatchSet. You use XssMatchSetId to get information about an XssMatchSet (see GetXssMatchSet), update an XssMatchSet (see UpdateXssMatchSet), insert an XssMatchSet into a Rule or delete one from a Rule (see UpdateRule), and delete an XssMatchSet from AWS WAF (see DeleteXssMatchSet).

XssMatchSetId is returned by CreateXssMatchSet and by ListXssMatchSets.

xssMatchSet_xssMatchTuples :: Lens' XssMatchSet [XssMatchTuple]

Specifies the parts of web requests that you want to inspect for cross-site scripting attacks.