Network.AWS.WAFRegional.Types.XssMatchSetSummary

Description

Documentation

data XssMatchSetSummary

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

The Id and Name of an XssMatchSet.

See: newXssMatchSetSummary smart constructor.

Constructors

XssMatchSetSummary'

Fields

xssMatchSetId :: Text

A unique identifier for an XssMatchSet. You use XssMatchSetId to get information about a XssMatchSet (see GetXssMatchSet), update an XssMatchSet (see UpdateXssMatchSet), insert an XssMatchSet into a Rule or delete one from a Rule (see UpdateRule), and delete an XssMatchSet from AWS WAF (see DeleteXssMatchSet).

XssMatchSetId is returned by CreateXssMatchSet and by ListXssMatchSets.

name :: Text

The name of the XssMatchSet, if any, specified by Id.

newXssMatchSetSummary

Arguments

:: Text
$sel:xssMatchSetId:XssMatchSetSummary'

-> Text
$sel:name:XssMatchSetSummary'

-> XssMatchSetSummary

Create a value of XssMatchSetSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:xssMatchSetId:XssMatchSetSummary', xssMatchSetSummary_xssMatchSetId - A unique identifier for an XssMatchSet. You use XssMatchSetId to get information about a XssMatchSet (see GetXssMatchSet), update an XssMatchSet (see UpdateXssMatchSet), insert an XssMatchSet into a Rule or delete one from a Rule (see UpdateRule), and delete an XssMatchSet from AWS WAF (see DeleteXssMatchSet).

XssMatchSetId is returned by CreateXssMatchSet and by ListXssMatchSets.

$sel:name:XssMatchSetSummary', xssMatchSetSummary_name - The name of the XssMatchSet, if any, specified by Id.

xssMatchSetSummary_xssMatchSetId :: Lens' XssMatchSetSummary Text

A unique identifier for an XssMatchSet. You use XssMatchSetId to get information about a XssMatchSet (see GetXssMatchSet), update an XssMatchSet (see UpdateXssMatchSet), insert an XssMatchSet into a Rule or delete one from a Rule (see UpdateRule), and delete an XssMatchSet from AWS WAF (see DeleteXssMatchSet).

XssMatchSetId is returned by CreateXssMatchSet and by ListXssMatchSets.

xssMatchSetSummary_name :: Lens' XssMatchSetSummary Text

The name of the XssMatchSet, if any, specified by Id.