Network.AWS.WAFRegional.UpdateRule

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Inserts or deletes Predicate objects in a Rule. Each Predicate object identifies a predicate, such as a ByteMatchSet or an IPSet, that specifies the web requests that you want to allow, block, or count. If you add more than one predicate to a Rule, a request must match all of the specifications to be allowed, blocked, or counted. For example, suppose that you add the following to a Rule:

A ByteMatchSet that matches the value BadBot in the User-Agent header

An IPSet that matches the IP address 192.0.2.44

You then add the Rule to a WebACL and specify that you want to block requests that satisfy the Rule. For a request to be blocked, the User-Agent header in the request must contain the value BadBot and the request must originate from the IP address 192.0.2.44.

To create and configure a Rule, perform the following steps:

Create and update the predicates that you want to include in the Rule.

Create the Rule. See CreateRule.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateRule request.

Submit an UpdateRule request to add predicates to the Rule.

Create and update a WebACL that contains the Rule. See CreateWebACL.

If you want to replace one ByteMatchSet or IPSet with another, you delete the existing one and add the new one.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide.

Creating a Request

data UpdateRule

See: newUpdateRule smart constructor.

Constructors

UpdateRule'

Fields

ruleId :: Text

The RuleId of the Rule that you want to update. RuleId is returned by CreateRule and by ListRules.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

updates :: [RuleUpdate]

An array of RuleUpdate objects that you want to insert into or delete from a Rule. For more information, see the applicable data types:

RuleUpdate: Contains Action and Predicate

Predicate: Contains DataId, Negated, and Type

FieldToMatch: Contains Data and Type

newUpdateRule

Arguments

:: Text
$sel:ruleId:UpdateRule'

-> Text
$sel:changeToken:UpdateRule'

-> UpdateRule

Create a value of UpdateRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleId:UpdateRule', updateRule_ruleId - The RuleId of the Rule that you want to update. RuleId is returned by CreateRule and by ListRules.

$sel:changeToken:UpdateRule', updateRule_changeToken - The value returned by the most recent call to GetChangeToken.

$sel:updates:UpdateRule', updateRule_updates - An array of RuleUpdate objects that you want to insert into or delete from a Rule. For more information, see the applicable data types:

RuleUpdate: Contains Action and Predicate

Predicate: Contains DataId, Negated, and Type

FieldToMatch: Contains Data and Type

Request Lenses

updateRule_ruleId :: Lens' UpdateRule Text

The RuleId of the Rule that you want to update. RuleId is returned by CreateRule and by ListRules.

updateRule_changeToken :: Lens' UpdateRule Text

The value returned by the most recent call to GetChangeToken.

updateRule_updates :: Lens' UpdateRule [RuleUpdate]

An array of RuleUpdate objects that you want to insert into or delete from a Rule. For more information, see the applicable data types:

RuleUpdate: Contains Action and Predicate

Predicate: Contains DataId, Negated, and Type

FieldToMatch: Contains Data and Type

Destructuring the Response

data UpdateRuleResponse

See: newUpdateRuleResponse smart constructor.

Constructors

UpdateRuleResponse'

Fields

changeToken :: Maybe Text

The ChangeToken that you used to submit the UpdateRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newUpdateRuleResponse

Arguments

:: Int
$sel:httpStatus:UpdateRuleResponse'

-> UpdateRuleResponse

Create a value of UpdateRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeToken:UpdateRule', updateRuleResponse_changeToken - The ChangeToken that you used to submit the UpdateRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:UpdateRuleResponse', updateRuleResponse_httpStatus - The response's http status code.

Response Lenses

updateRuleResponse_changeToken :: Lens' UpdateRuleResponse (Maybe Text)

The ChangeToken that you used to submit the UpdateRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

updateRuleResponse_httpStatus :: Lens' UpdateRuleResponse Int

The response's http status code.