Network.AWS.WellArchitected.CreateWorkload

Description

Create a new workload.

The owner of a workload can share the workload with other AWS accounts and IAM users in the same AWS Region. Only the owner of a workload can delete it.

For more information, see Defining a Workload in the AWS Well-Architected Tool User Guide.

Creating a Request

data CreateWorkload

Input for workload creation.

See: newCreateWorkload smart constructor.

Constructors

CreateWorkload'

Fields

architecturalDesign :: Maybe Text

accountIds :: Maybe [Text]

industry :: Maybe Text

awsRegions :: Maybe [Text]

industryType :: Maybe Text

notes :: Maybe Text

pillarPriorities :: Maybe [Text]

nonAwsRegions :: Maybe [Text]

tags :: Maybe (HashMap Text Text)

The tags to be associated with the workload.

workloadName :: Text

description :: Text

environment :: WorkloadEnvironment

reviewOwner :: Text

lenses :: [Text]

clientRequestToken :: Text

newCreateWorkload

Arguments

:: Text
$sel:workloadName:CreateWorkload'

-> Text
$sel:description:CreateWorkload'

-> WorkloadEnvironment
$sel:environment:CreateWorkload'

-> Text
$sel:reviewOwner:CreateWorkload'

-> Text
$sel:clientRequestToken:CreateWorkload'

-> CreateWorkload

Create a value of CreateWorkload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:architecturalDesign:CreateWorkload', createWorkload_architecturalDesign - Undocumented member.

$sel:accountIds:CreateWorkload', createWorkload_accountIds - Undocumented member.

$sel:industry:CreateWorkload', createWorkload_industry - Undocumented member.

$sel:awsRegions:CreateWorkload', createWorkload_awsRegions - Undocumented member.

$sel:industryType:CreateWorkload', createWorkload_industryType - Undocumented member.

$sel:notes:CreateWorkload', createWorkload_notes - Undocumented member.

$sel:pillarPriorities:CreateWorkload', createWorkload_pillarPriorities - Undocumented member.

$sel:nonAwsRegions:CreateWorkload', createWorkload_nonAwsRegions - Undocumented member.

$sel:tags:CreateWorkload', createWorkload_tags - The tags to be associated with the workload.

$sel:workloadName:CreateWorkload', createWorkload_workloadName - Undocumented member.

$sel:description:CreateWorkload', createWorkload_description - Undocumented member.

$sel:environment:CreateWorkload', createWorkload_environment - Undocumented member.

$sel:reviewOwner:CreateWorkload', createWorkload_reviewOwner - Undocumented member.

$sel:lenses:CreateWorkload', createWorkload_lenses - Undocumented member.

$sel:clientRequestToken:CreateWorkload', createWorkload_clientRequestToken - Undocumented member.

Request Lenses

createWorkload_architecturalDesign :: Lens' CreateWorkload (Maybe Text)

Undocumented member.

createWorkload_accountIds :: Lens' CreateWorkload (Maybe [Text])

Undocumented member.

createWorkload_industry :: Lens' CreateWorkload (Maybe Text)

Undocumented member.

createWorkload_awsRegions :: Lens' CreateWorkload (Maybe [Text])

Undocumented member.

createWorkload_industryType :: Lens' CreateWorkload (Maybe Text)

Undocumented member.

createWorkload_notes :: Lens' CreateWorkload (Maybe Text)

Undocumented member.

createWorkload_pillarPriorities :: Lens' CreateWorkload (Maybe [Text])

Undocumented member.

createWorkload_nonAwsRegions :: Lens' CreateWorkload (Maybe [Text])

Undocumented member.

createWorkload_tags :: Lens' CreateWorkload (Maybe (HashMap Text Text))

The tags to be associated with the workload.

createWorkload_workloadName :: Lens' CreateWorkload Text

Undocumented member.

createWorkload_description :: Lens' CreateWorkload Text

Undocumented member.

createWorkload_environment :: Lens' CreateWorkload WorkloadEnvironment

Undocumented member.

createWorkload_reviewOwner :: Lens' CreateWorkload Text

Undocumented member.

createWorkload_lenses :: Lens' CreateWorkload [Text]

Undocumented member.

createWorkload_clientRequestToken :: Lens' CreateWorkload Text

Undocumented member.

Destructuring the Response

data CreateWorkloadResponse

Output of a create workload call.

See: newCreateWorkloadResponse smart constructor.

Constructors

CreateWorkloadResponse'

Fields

workloadArn :: Maybe Text

workloadId :: Maybe Text

httpStatus :: Int

The response's http status code.

newCreateWorkloadResponse

Arguments

:: Int
$sel:httpStatus:CreateWorkloadResponse'

-> CreateWorkloadResponse

Create a value of CreateWorkloadResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workloadArn:CreateWorkloadResponse', createWorkloadResponse_workloadArn - Undocumented member.

$sel:workloadId:CreateWorkloadResponse', createWorkloadResponse_workloadId - Undocumented member.

$sel:httpStatus:CreateWorkloadResponse', createWorkloadResponse_httpStatus - The response's http status code.

Response Lenses

createWorkloadResponse_workloadArn :: Lens' CreateWorkloadResponse (Maybe Text)

Undocumented member.

createWorkloadResponse_workloadId :: Lens' CreateWorkloadResponse (Maybe Text)

Undocumented member.

createWorkloadResponse_httpStatus :: Lens' CreateWorkloadResponse Int

The response's http status code.