Network.AWS.WellArchitected.GetMilestone

Description

Get a milestone for an existing workload.

Creating a Request

data GetMilestone

Input to get a milestone.

See: newGetMilestone smart constructor.

Constructors

GetMilestone'

Fields

workloadId :: Text

milestoneNumber :: Natural

newGetMilestone

Arguments

:: Text
$sel:workloadId:GetMilestone'

-> Natural
$sel:milestoneNumber:GetMilestone'

-> GetMilestone

Create a value of GetMilestone with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workloadId:GetMilestone', getMilestone_workloadId - Undocumented member.

$sel:milestoneNumber:GetMilestone', getMilestone_milestoneNumber - Undocumented member.

Request Lenses

getMilestone_workloadId :: Lens' GetMilestone Text

Undocumented member.

getMilestone_milestoneNumber :: Lens' GetMilestone Natural

Undocumented member.

Destructuring the Response

data GetMilestoneResponse

Output of a get milestone call.

See: newGetMilestoneResponse smart constructor.

Constructors

GetMilestoneResponse'

Fields

milestone :: Maybe Milestone

workloadId :: Maybe Text

httpStatus :: Int

The response's http status code.

newGetMilestoneResponse

Arguments

:: Int
$sel:httpStatus:GetMilestoneResponse'

-> GetMilestoneResponse

Create a value of GetMilestoneResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:milestone:GetMilestoneResponse', getMilestoneResponse_milestone - Undocumented member.

$sel:workloadId:GetMilestone', getMilestoneResponse_workloadId - Undocumented member.

$sel:httpStatus:GetMilestoneResponse', getMilestoneResponse_httpStatus - The response's http status code.

Response Lenses

getMilestoneResponse_milestone :: Lens' GetMilestoneResponse (Maybe Milestone)

Undocumented member.

getMilestoneResponse_workloadId :: Lens' GetMilestoneResponse (Maybe Text)

Undocumented member.

getMilestoneResponse_httpStatus :: Lens' GetMilestoneResponse Int

The response's http status code.