Network.AWS.WellArchitected.Types.Answer

Description

Documentation

data Answer

An answer of the question.

See: newAnswer smart constructor.

Constructors

Answer'

Fields

choiceAnswers :: Maybe [ChoiceAnswer]

A list of selected choices to a question in your workload.

helpfulResourceUrl :: Maybe Text

isApplicable :: Maybe Bool

pillarId :: Maybe Text

improvementPlanUrl :: Maybe Text

questionDescription :: Maybe Text

risk :: Maybe Risk

questionTitle :: Maybe Text

selectedChoices :: Maybe [Text]

reason :: Maybe AnswerReason

The reason why the question is not applicable to your workload.

choices :: Maybe [Choice]

questionId :: Maybe Text

notes :: Maybe Text

newAnswer :: Answer

Create a value of Answer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:choiceAnswers:Answer', answer_choiceAnswers - A list of selected choices to a question in your workload.

$sel:helpfulResourceUrl:Answer', answer_helpfulResourceUrl - Undocumented member.

$sel:isApplicable:Answer', answer_isApplicable - Undocumented member.

$sel:pillarId:Answer', answer_pillarId - Undocumented member.

$sel:improvementPlanUrl:Answer', answer_improvementPlanUrl - Undocumented member.

$sel:questionDescription:Answer', answer_questionDescription - Undocumented member.

$sel:risk:Answer', answer_risk - Undocumented member.

$sel:questionTitle:Answer', answer_questionTitle - Undocumented member.

$sel:selectedChoices:Answer', answer_selectedChoices - Undocumented member.

$sel:reason:Answer', answer_reason - The reason why the question is not applicable to your workload.

$sel:choices:Answer', answer_choices - Undocumented member.

$sel:questionId:Answer', answer_questionId - Undocumented member.

$sel:notes:Answer', answer_notes - Undocumented member.

answer_choiceAnswers :: Lens' Answer (Maybe [ChoiceAnswer])

A list of selected choices to a question in your workload.

answer_helpfulResourceUrl :: Lens' Answer (Maybe Text)

Undocumented member.

answer_isApplicable :: Lens' Answer (Maybe Bool)

Undocumented member.

answer_pillarId :: Lens' Answer (Maybe Text)

Undocumented member.

answer_improvementPlanUrl :: Lens' Answer (Maybe Text)

Undocumented member.

answer_questionDescription :: Lens' Answer (Maybe Text)

Undocumented member.

answer_risk :: Lens' Answer (Maybe Risk)

Undocumented member.

answer_questionTitle :: Lens' Answer (Maybe Text)

Undocumented member.

answer_selectedChoices :: Lens' Answer (Maybe [Text])

Undocumented member.

answer_reason :: Lens' Answer (Maybe AnswerReason)

The reason why the question is not applicable to your workload.

answer_choices :: Lens' Answer (Maybe [Choice])

Undocumented member.

answer_questionId :: Lens' Answer (Maybe Text)

Undocumented member.

answer_notes :: Lens' Answer (Maybe Text)

Undocumented member.