Network.AWS.WellArchitected.Types.AnswerSummary

Description

Documentation

data AnswerSummary

An answer summary of a lens review in a workload.

See: newAnswerSummary smart constructor.

Constructors

AnswerSummary'

Fields

isApplicable :: Maybe Bool

pillarId :: Maybe Text

choiceAnswerSummaries :: Maybe [ChoiceAnswerSummary]

A list of selected choices to a question in your workload.

risk :: Maybe Risk

questionTitle :: Maybe Text

selectedChoices :: Maybe [Text]

reason :: Maybe AnswerReason

The reason why a choice is non-applicable to a question in your workload.

choices :: Maybe [Choice]

questionId :: Maybe Text

newAnswerSummary :: AnswerSummary

Create a value of AnswerSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isApplicable:AnswerSummary', answerSummary_isApplicable - Undocumented member.

$sel:pillarId:AnswerSummary', answerSummary_pillarId - Undocumented member.

$sel:choiceAnswerSummaries:AnswerSummary', answerSummary_choiceAnswerSummaries - A list of selected choices to a question in your workload.

$sel:risk:AnswerSummary', answerSummary_risk - Undocumented member.

$sel:questionTitle:AnswerSummary', answerSummary_questionTitle - Undocumented member.

$sel:selectedChoices:AnswerSummary', answerSummary_selectedChoices - Undocumented member.

$sel:reason:AnswerSummary', answerSummary_reason - The reason why a choice is non-applicable to a question in your workload.

$sel:choices:AnswerSummary', answerSummary_choices - Undocumented member.

$sel:questionId:AnswerSummary', answerSummary_questionId - Undocumented member.

answerSummary_isApplicable :: Lens' AnswerSummary (Maybe Bool)

Undocumented member.

answerSummary_pillarId :: Lens' AnswerSummary (Maybe Text)

Undocumented member.

answerSummary_choiceAnswerSummaries :: Lens' AnswerSummary (Maybe [ChoiceAnswerSummary])

A list of selected choices to a question in your workload.

answerSummary_risk :: Lens' AnswerSummary (Maybe Risk)

Undocumented member.

answerSummary_questionTitle :: Lens' AnswerSummary (Maybe Text)

Undocumented member.

answerSummary_selectedChoices :: Lens' AnswerSummary (Maybe [Text])

Undocumented member.

answerSummary_reason :: Lens' AnswerSummary (Maybe AnswerReason)

The reason why a choice is non-applicable to a question in your workload.

answerSummary_choices :: Lens' AnswerSummary (Maybe [Choice])

Undocumented member.

answerSummary_questionId :: Lens' AnswerSummary (Maybe Text)

Undocumented member.