Network.AWS.WellArchitected.Types.Milestone

Description

Documentation

data Milestone

A milestone return object.

See: newMilestone smart constructor.

Constructors

Milestone'

Fields

workload :: Maybe Workload

milestoneNumber :: Maybe Natural

milestoneName :: Maybe Text

recordedAt :: Maybe POSIX

newMilestone :: Milestone

Create a value of Milestone with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workload:Milestone', milestone_workload - Undocumented member.

$sel:milestoneNumber:Milestone', milestone_milestoneNumber - Undocumented member.

$sel:milestoneName:Milestone', milestone_milestoneName - Undocumented member.

$sel:recordedAt:Milestone', milestone_recordedAt - Undocumented member.

milestone_workload :: Lens' Milestone (Maybe Workload)

Undocumented member.

milestone_milestoneNumber :: Lens' Milestone (Maybe Natural)

Undocumented member.

milestone_milestoneName :: Lens' Milestone (Maybe Text)

Undocumented member.

milestone_recordedAt :: Lens' Milestone (Maybe UTCTime)

Undocumented member.