Network.AWS.WellArchitected.UpdateWorkload

Description

Update an existing workload.

Creating a Request

data UpdateWorkload

Input to update a workload.

See: newUpdateWorkload smart constructor.

Constructors

UpdateWorkload'

Fields

isReviewOwnerUpdateAcknowledged :: Maybe Bool

Flag indicating whether the workload owner has acknowledged that the Review owner field is required.

If a Review owner is not added to the workload within 60 days of acknowledgement, access to the workload is restricted until an owner is added.

architecturalDesign :: Maybe Text

accountIds :: Maybe [Text]

industry :: Maybe Text

environment :: Maybe WorkloadEnvironment

awsRegions :: Maybe [Text]

improvementStatus :: Maybe WorkloadImprovementStatus

industryType :: Maybe Text

workloadName :: Maybe Text

notes :: Maybe Text

reviewOwner :: Maybe Text

description :: Maybe Text

pillarPriorities :: Maybe [Text]

nonAwsRegions :: Maybe [Text]

workloadId :: Text

newUpdateWorkload

Arguments

:: Text
$sel:workloadId:UpdateWorkload'

-> UpdateWorkload

Create a value of UpdateWorkload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isReviewOwnerUpdateAcknowledged:UpdateWorkload', updateWorkload_isReviewOwnerUpdateAcknowledged - Flag indicating whether the workload owner has acknowledged that the Review owner field is required.

If a Review owner is not added to the workload within 60 days of acknowledgement, access to the workload is restricted until an owner is added.

$sel:architecturalDesign:UpdateWorkload', updateWorkload_architecturalDesign - Undocumented member.

$sel:accountIds:UpdateWorkload', updateWorkload_accountIds - Undocumented member.

$sel:industry:UpdateWorkload', updateWorkload_industry - Undocumented member.

$sel:environment:UpdateWorkload', updateWorkload_environment - Undocumented member.

$sel:awsRegions:UpdateWorkload', updateWorkload_awsRegions - Undocumented member.

$sel:improvementStatus:UpdateWorkload', updateWorkload_improvementStatus - Undocumented member.

$sel:industryType:UpdateWorkload', updateWorkload_industryType - Undocumented member.

$sel:workloadName:UpdateWorkload', updateWorkload_workloadName - Undocumented member.

$sel:notes:UpdateWorkload', updateWorkload_notes - Undocumented member.

$sel:reviewOwner:UpdateWorkload', updateWorkload_reviewOwner - Undocumented member.

$sel:description:UpdateWorkload', updateWorkload_description - Undocumented member.

$sel:pillarPriorities:UpdateWorkload', updateWorkload_pillarPriorities - Undocumented member.

$sel:nonAwsRegions:UpdateWorkload', updateWorkload_nonAwsRegions - Undocumented member.

$sel:workloadId:UpdateWorkload', updateWorkload_workloadId - Undocumented member.

Request Lenses

updateWorkload_isReviewOwnerUpdateAcknowledged :: Lens' UpdateWorkload (Maybe Bool)

Flag indicating whether the workload owner has acknowledged that the Review owner field is required.

If a Review owner is not added to the workload within 60 days of acknowledgement, access to the workload is restricted until an owner is added.

updateWorkload_architecturalDesign :: Lens' UpdateWorkload (Maybe Text)

Undocumented member.

updateWorkload_accountIds :: Lens' UpdateWorkload (Maybe [Text])

Undocumented member.

updateWorkload_industry :: Lens' UpdateWorkload (Maybe Text)

Undocumented member.

updateWorkload_environment :: Lens' UpdateWorkload (Maybe WorkloadEnvironment)

Undocumented member.

updateWorkload_awsRegions :: Lens' UpdateWorkload (Maybe [Text])

Undocumented member.

updateWorkload_improvementStatus :: Lens' UpdateWorkload (Maybe WorkloadImprovementStatus)

Undocumented member.

updateWorkload_industryType :: Lens' UpdateWorkload (Maybe Text)

Undocumented member.

updateWorkload_workloadName :: Lens' UpdateWorkload (Maybe Text)

Undocumented member.

updateWorkload_notes :: Lens' UpdateWorkload (Maybe Text)

Undocumented member.

updateWorkload_reviewOwner :: Lens' UpdateWorkload (Maybe Text)

Undocumented member.

updateWorkload_description :: Lens' UpdateWorkload (Maybe Text)

Undocumented member.

updateWorkload_pillarPriorities :: Lens' UpdateWorkload (Maybe [Text])

Undocumented member.

updateWorkload_nonAwsRegions :: Lens' UpdateWorkload (Maybe [Text])

Undocumented member.

updateWorkload_workloadId :: Lens' UpdateWorkload Text

Undocumented member.

Destructuring the Response

data UpdateWorkloadResponse

Output of an update workload call.

See: newUpdateWorkloadResponse smart constructor.

Constructors

UpdateWorkloadResponse'

Fields

workload :: Maybe Workload

httpStatus :: Int

The response's http status code.

newUpdateWorkloadResponse

Arguments

:: Int
$sel:httpStatus:UpdateWorkloadResponse'

-> UpdateWorkloadResponse

Create a value of UpdateWorkloadResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workload:UpdateWorkloadResponse', updateWorkloadResponse_workload - Undocumented member.

$sel:httpStatus:UpdateWorkloadResponse', updateWorkloadResponse_httpStatus - The response's http status code.

Response Lenses

updateWorkloadResponse_workload :: Lens' UpdateWorkloadResponse (Maybe Workload)

Undocumented member.

updateWorkloadResponse_httpStatus :: Lens' UpdateWorkloadResponse Int

The response's http status code.