Network.AWS.Wisdom.CreateContent

Description

Creates Wisdom content. Before to calling this API, use StartContentUpload to upload an asset.

Creating a Request

data CreateContent

See: newCreateContent smart constructor.

Constructors

CreateContent'

Fields

overrideLinkOutUri :: Maybe Text

The URI you want to use for the article. If the knowledge base has a templateUri, setting this argument overrides it for this piece of content.

clientToken :: Maybe Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

metadata :: Maybe (HashMap Text Text)

A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

title :: Maybe Text

The title of the content. If not set, the title is equal to the name.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

knowledgeBaseId :: Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

name :: Text

The name of the content. Each piece of content in a knowledge base must have a unique name. You can retrieve a piece of content using only its knowledge base and its name with the SearchContent API.

uploadId :: Text

A pointer to the uploaded asset. This value is returned by StartContentUpload.

newCreateContent

Arguments

:: Text
$sel:knowledgeBaseId:CreateContent'

-> Text
$sel:name:CreateContent'

-> Text
$sel:uploadId:CreateContent'

-> CreateContent

Create a value of CreateContent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:overrideLinkOutUri:CreateContent', createContent_overrideLinkOutUri - The URI you want to use for the article. If the knowledge base has a templateUri, setting this argument overrides it for this piece of content.

$sel:clientToken:CreateContent', createContent_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:metadata:CreateContent', createContent_metadata - A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

$sel:title:CreateContent', createContent_title - The title of the content. If not set, the title is equal to the name.

$sel:tags:CreateContent', createContent_tags - The tags used to organize, track, or control access for this resource.

$sel:knowledgeBaseId:CreateContent', createContent_knowledgeBaseId - The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

$sel:name:CreateContent', createContent_name - The name of the content. Each piece of content in a knowledge base must have a unique name. You can retrieve a piece of content using only its knowledge base and its name with the SearchContent API.

$sel:uploadId:CreateContent', createContent_uploadId - A pointer to the uploaded asset. This value is returned by StartContentUpload.

Request Lenses

createContent_overrideLinkOutUri :: Lens' CreateContent (Maybe Text)

The URI you want to use for the article. If the knowledge base has a templateUri, setting this argument overrides it for this piece of content.

createContent_clientToken :: Lens' CreateContent (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createContent_metadata :: Lens' CreateContent (Maybe (HashMap Text Text))

A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

createContent_title :: Lens' CreateContent (Maybe Text)

The title of the content. If not set, the title is equal to the name.

createContent_tags :: Lens' CreateContent (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

createContent_knowledgeBaseId :: Lens' CreateContent Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

createContent_name :: Lens' CreateContent Text

The name of the content. Each piece of content in a knowledge base must have a unique name. You can retrieve a piece of content using only its knowledge base and its name with the SearchContent API.

createContent_uploadId :: Lens' CreateContent Text

A pointer to the uploaded asset. This value is returned by StartContentUpload.

Destructuring the Response

data CreateContentResponse

See: newCreateContentResponse smart constructor.

Constructors

CreateContentResponse'

Fields

content :: Maybe ContentData

The content.

httpStatus :: Int

The response's http status code.

newCreateContentResponse

Arguments

:: Int
$sel:httpStatus:CreateContentResponse'

-> CreateContentResponse

Create a value of CreateContentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:content:CreateContentResponse', createContentResponse_content - The content.

$sel:httpStatus:CreateContentResponse', createContentResponse_httpStatus - The response's http status code.

Response Lenses

createContentResponse_content :: Lens' CreateContentResponse (Maybe ContentData)

The content.

createContentResponse_httpStatus :: Lens' CreateContentResponse Int

The response's http status code.