Network.AWS.Wisdom.CreateKnowledgeBase

Description

Creates a knowledge base.

When using this API, you cannot reuse Amazon AppIntegrations DataIntegrations with external knowledge bases such as Salesforce and ServiceNow. If you do, you'll get an InvalidRequestException error.

 <p>For example, you're programmatically managing your external knowledge base, and you want to add or remove one of the fields that is being ingested from Salesforce. Do the following:</p> <ol> <li> <p>Call <a href="https://docs.aws.amazon.com/wisdom/latest/APIReference/API_DeleteKnowledgeBase.html">DeleteKnowledgeBase</a>.</p> </li> <li> <p>Call <a href="https://docs.aws.amazon.com/appintegrations/latest/APIReference/API_DeleteDataIntegration.html">DeleteDataIntegration</a>.</p> </li> <li> <p>Call <a href="https://docs.aws.amazon.com/appintegrations/latest/APIReference/API_CreateDataIntegration.html">CreateDataIntegration</a> to recreate the DataIntegration or a create different one.</p> </li> <li> <p>Call CreateKnowledgeBase.</p> </li> </ol> </note>

Creating a Request

data CreateKnowledgeBase

See: newCreateKnowledgeBase smart constructor.

Constructors

CreateKnowledgeBase'

Fields

clientToken :: Maybe Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

renderingConfiguration :: Maybe RenderingConfiguration

Information about how to render the content.

sourceConfiguration :: Maybe SourceConfiguration

The source of the knowledge base content. Only set this argument for EXTERNAL knowledge bases.

description :: Maybe Text

The description.

serverSideEncryptionConfiguration :: Maybe ServerSideEncryptionConfiguration

The KMS key used for encryption.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

knowledgeBaseType :: KnowledgeBaseType

The type of knowledge base. Only CUSTOM knowledge bases allow you to upload your own content. EXTERNAL knowledge bases support integrations with third-party systems whose content is synchronized automatically.

name :: Text

The name of the knowledge base.

newCreateKnowledgeBase

Arguments

:: KnowledgeBaseType
$sel:knowledgeBaseType:CreateKnowledgeBase'

-> Text
$sel:name:CreateKnowledgeBase'

-> CreateKnowledgeBase

Create a value of CreateKnowledgeBase with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateKnowledgeBase', createKnowledgeBase_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:renderingConfiguration:CreateKnowledgeBase', createKnowledgeBase_renderingConfiguration - Information about how to render the content.

$sel:sourceConfiguration:CreateKnowledgeBase', createKnowledgeBase_sourceConfiguration - The source of the knowledge base content. Only set this argument for EXTERNAL knowledge bases.

$sel:description:CreateKnowledgeBase', createKnowledgeBase_description - The description.

$sel:serverSideEncryptionConfiguration:CreateKnowledgeBase', createKnowledgeBase_serverSideEncryptionConfiguration - The KMS key used for encryption.

$sel:tags:CreateKnowledgeBase', createKnowledgeBase_tags - The tags used to organize, track, or control access for this resource.

$sel:knowledgeBaseType:CreateKnowledgeBase', createKnowledgeBase_knowledgeBaseType - The type of knowledge base. Only CUSTOM knowledge bases allow you to upload your own content. EXTERNAL knowledge bases support integrations with third-party systems whose content is synchronized automatically.

$sel:name:CreateKnowledgeBase', createKnowledgeBase_name - The name of the knowledge base.

Request Lenses

createKnowledgeBase_clientToken :: Lens' CreateKnowledgeBase (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createKnowledgeBase_renderingConfiguration :: Lens' CreateKnowledgeBase (Maybe RenderingConfiguration)

Information about how to render the content.

createKnowledgeBase_sourceConfiguration :: Lens' CreateKnowledgeBase (Maybe SourceConfiguration)

The source of the knowledge base content. Only set this argument for EXTERNAL knowledge bases.

createKnowledgeBase_description :: Lens' CreateKnowledgeBase (Maybe Text)

The description.

createKnowledgeBase_serverSideEncryptionConfiguration :: Lens' CreateKnowledgeBase (Maybe ServerSideEncryptionConfiguration)

The KMS key used for encryption.

createKnowledgeBase_tags :: Lens' CreateKnowledgeBase (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

createKnowledgeBase_knowledgeBaseType :: Lens' CreateKnowledgeBase KnowledgeBaseType

The type of knowledge base. Only CUSTOM knowledge bases allow you to upload your own content. EXTERNAL knowledge bases support integrations with third-party systems whose content is synchronized automatically.

createKnowledgeBase_name :: Lens' CreateKnowledgeBase Text

The name of the knowledge base.

Destructuring the Response

data CreateKnowledgeBaseResponse

See: newCreateKnowledgeBaseResponse smart constructor.

Constructors

CreateKnowledgeBaseResponse'

Fields

knowledgeBase :: Maybe KnowledgeBaseData

The knowledge base.

httpStatus :: Int

The response's http status code.

newCreateKnowledgeBaseResponse

Arguments

:: Int
$sel:httpStatus:CreateKnowledgeBaseResponse'

-> CreateKnowledgeBaseResponse

Create a value of CreateKnowledgeBaseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:knowledgeBase:CreateKnowledgeBaseResponse', createKnowledgeBaseResponse_knowledgeBase - The knowledge base.

$sel:httpStatus:CreateKnowledgeBaseResponse', createKnowledgeBaseResponse_httpStatus - The response's http status code.

Response Lenses

createKnowledgeBaseResponse_knowledgeBase :: Lens' CreateKnowledgeBaseResponse (Maybe KnowledgeBaseData)

The knowledge base.

createKnowledgeBaseResponse_httpStatus :: Lens' CreateKnowledgeBaseResponse Int

The response's http status code.