Network.AWS.Wisdom.SearchSessions

Description

Searches for sessions.

This operation returns paginated results.

Creating a Request

data SearchSessions

See: newSearchSessions smart constructor.

Constructors

SearchSessions'

Fields

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return per page.

assistantId :: Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

searchExpression :: SearchExpression

The search expression to filter results.

newSearchSessions

Arguments

:: Text
$sel:assistantId:SearchSessions'

-> SearchExpression
$sel:searchExpression:SearchSessions'

-> SearchSessions

Create a value of SearchSessions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchSessions', searchSessions_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:maxResults:SearchSessions', searchSessions_maxResults - The maximum number of results to return per page.

$sel:assistantId:SearchSessions', searchSessions_assistantId - The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

$sel:searchExpression:SearchSessions', searchSessions_searchExpression - The search expression to filter results.

Request Lenses

searchSessions_nextToken :: Lens' SearchSessions (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

searchSessions_maxResults :: Lens' SearchSessions (Maybe Natural)

The maximum number of results to return per page.

searchSessions_assistantId :: Lens' SearchSessions Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

searchSessions_searchExpression :: Lens' SearchSessions SearchExpression

The search expression to filter results.

Destructuring the Response

data SearchSessionsResponse

See: newSearchSessionsResponse smart constructor.

Constructors

SearchSessionsResponse'

Fields

nextToken :: Maybe Text

If there are additional results, this is the token for the next set of results.

httpStatus :: Int

The response's http status code.

sessionSummaries :: [SessionSummary]

Summary information about the sessions.

newSearchSessionsResponse

Arguments

:: Int
$sel:httpStatus:SearchSessionsResponse'

-> SearchSessionsResponse

Create a value of SearchSessionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchSessions', searchSessionsResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:httpStatus:SearchSessionsResponse', searchSessionsResponse_httpStatus - The response's http status code.

$sel:sessionSummaries:SearchSessionsResponse', searchSessionsResponse_sessionSummaries - Summary information about the sessions.

Response Lenses

searchSessionsResponse_nextToken :: Lens' SearchSessionsResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

searchSessionsResponse_httpStatus :: Lens' SearchSessionsResponse Int

The response's http status code.

searchSessionsResponse_sessionSummaries :: Lens' SearchSessionsResponse [SessionSummary]

Summary information about the sessions.