Network.AWS.Wisdom.StartContentUpload

Description

Get a URL to upload content to a knowledge base. To upload content, first make a PUT request to the returned URL with your file, making sure to include the required headers. Then use CreateContent to finalize the content creation process or UpdateContent to modify an existing resource. You can only upload content to a knowledge base of type CUSTOM.

Creating a Request

data StartContentUpload

See: newStartContentUpload smart constructor.

Constructors

StartContentUpload'

Fields

contentType :: Text

The type of content to upload.

knowledgeBaseId :: Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

newStartContentUpload

Arguments

:: Text
$sel:contentType:StartContentUpload'

-> Text
$sel:knowledgeBaseId:StartContentUpload'

-> StartContentUpload

Create a value of StartContentUpload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentType:StartContentUpload', startContentUpload_contentType - The type of content to upload.

$sel:knowledgeBaseId:StartContentUpload', startContentUpload_knowledgeBaseId - The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

Request Lenses

startContentUpload_contentType :: Lens' StartContentUpload Text

The type of content to upload.

startContentUpload_knowledgeBaseId :: Lens' StartContentUpload Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

Destructuring the Response

data StartContentUploadResponse

See: newStartContentUploadResponse smart constructor.

Constructors

StartContentUploadResponse'

Fields

httpStatus :: Int

The response's http status code.

headersToInclude :: HashMap Text Text

The headers to include in the upload.

uploadId :: Text

The identifier of the upload.

url :: Sensitive Text

The URL of the upload.

urlExpiry :: POSIX

The expiration time of the URL as an epoch timestamp.

newStartContentUploadResponse

Arguments

:: Int
$sel:httpStatus:StartContentUploadResponse'

-> Text
$sel:uploadId:StartContentUploadResponse'

-> Text
$sel:url:StartContentUploadResponse'

-> UTCTime
$sel:urlExpiry:StartContentUploadResponse'

-> StartContentUploadResponse

Create a value of StartContentUploadResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StartContentUploadResponse', startContentUploadResponse_httpStatus - The response's http status code.

$sel:headersToInclude:StartContentUploadResponse', startContentUploadResponse_headersToInclude - The headers to include in the upload.

$sel:uploadId:StartContentUploadResponse', startContentUploadResponse_uploadId - The identifier of the upload.

$sel:url:StartContentUploadResponse', startContentUploadResponse_url - The URL of the upload.

$sel:urlExpiry:StartContentUploadResponse', startContentUploadResponse_urlExpiry - The expiration time of the URL as an epoch timestamp.

Response Lenses

startContentUploadResponse_httpStatus :: Lens' StartContentUploadResponse Int

The response's http status code.

startContentUploadResponse_headersToInclude :: Lens' StartContentUploadResponse (HashMap Text Text)

The headers to include in the upload.

startContentUploadResponse_uploadId :: Lens' StartContentUploadResponse Text

The identifier of the upload.

startContentUploadResponse_url :: Lens' StartContentUploadResponse Text

The URL of the upload.

startContentUploadResponse_urlExpiry :: Lens' StartContentUploadResponse UTCTime

The expiration time of the URL as an epoch timestamp.