Network.AWS.Wisdom.Types.AssistantAssociationSummary

Description

Documentation

data AssistantAssociationSummary

Summary information about the assistant association.

See: newAssistantAssociationSummary smart constructor.

Constructors

AssistantAssociationSummary'

Fields

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

assistantArn :: Text

The Amazon Resource Name (ARN) of the Wisdom assistant

assistantAssociationArn :: Text

The Amazon Resource Name (ARN) of the assistant association.

assistantAssociationId :: Text

The identifier of the assistant association.

assistantId :: Text

The identifier of the Wisdom assistant.

associationData :: AssistantAssociationOutputData

The association data.

associationType :: AssociationType

The type of association.

newAssistantAssociationSummary

Arguments

:: Text
$sel:assistantArn:AssistantAssociationSummary'

-> Text
$sel:assistantAssociationArn:AssistantAssociationSummary'

-> Text
$sel:assistantAssociationId:AssistantAssociationSummary'

-> Text
$sel:assistantId:AssistantAssociationSummary'

-> AssistantAssociationOutputData
$sel:associationData:AssistantAssociationSummary'

-> AssociationType
$sel:associationType:AssistantAssociationSummary'

-> AssistantAssociationSummary

Create a value of AssistantAssociationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:AssistantAssociationSummary', assistantAssociationSummary_tags - The tags used to organize, track, or control access for this resource.

$sel:assistantArn:AssistantAssociationSummary', assistantAssociationSummary_assistantArn - The Amazon Resource Name (ARN) of the Wisdom assistant

$sel:assistantAssociationArn:AssistantAssociationSummary', assistantAssociationSummary_assistantAssociationArn - The Amazon Resource Name (ARN) of the assistant association.

$sel:assistantAssociationId:AssistantAssociationSummary', assistantAssociationSummary_assistantAssociationId - The identifier of the assistant association.

$sel:assistantId:AssistantAssociationSummary', assistantAssociationSummary_assistantId - The identifier of the Wisdom assistant.

$sel:associationData:AssistantAssociationSummary', assistantAssociationSummary_associationData - The association data.

$sel:associationType:AssistantAssociationSummary', assistantAssociationSummary_associationType - The type of association.

assistantAssociationSummary_tags :: Lens' AssistantAssociationSummary (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

assistantAssociationSummary_assistantArn :: Lens' AssistantAssociationSummary Text

The Amazon Resource Name (ARN) of the Wisdom assistant

assistantAssociationSummary_assistantAssociationArn :: Lens' AssistantAssociationSummary Text

The Amazon Resource Name (ARN) of the assistant association.

assistantAssociationSummary_assistantAssociationId :: Lens' AssistantAssociationSummary Text

The identifier of the assistant association.

assistantAssociationSummary_assistantId :: Lens' AssistantAssociationSummary Text

The identifier of the Wisdom assistant.

assistantAssociationSummary_associationData :: Lens' AssistantAssociationSummary AssistantAssociationOutputData

The association data.

assistantAssociationSummary_associationType :: Lens' AssistantAssociationSummary AssociationType

The type of association.