Network.AWS.Wisdom.Types.ContentSummary

Description

Documentation

data ContentSummary

Summary information about the content.

See: newContentSummary smart constructor.

Constructors

ContentSummary'

Fields

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

contentArn :: Text

The Amazon Resource Name (ARN) of the content.

contentId :: Text

The identifier of the content.

contentType :: Text

The media type of the content.

knowledgeBaseArn :: Text

The Amazon Resource Name (ARN) of the knowledge base.

knowledgeBaseId :: Text

The the identifier of the knowledge base.

metadata :: HashMap Text Text

A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

name :: Text

The name of the content.

revisionId :: Text

The identifier of the revision of the content.

status :: ContentStatus

The status of the content.

title :: Text

The title of the content.

newContentSummary

Arguments

:: Text
$sel:contentArn:ContentSummary'

-> Text
$sel:contentId:ContentSummary'

-> Text
$sel:contentType:ContentSummary'

-> Text
$sel:knowledgeBaseArn:ContentSummary'

-> Text
$sel:knowledgeBaseId:ContentSummary'

-> Text
$sel:name:ContentSummary'

-> Text
$sel:revisionId:ContentSummary'

-> ContentStatus
$sel:status:ContentSummary'

-> Text
$sel:title:ContentSummary'

-> ContentSummary

Create a value of ContentSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ContentSummary', contentSummary_tags - The tags used to organize, track, or control access for this resource.

$sel:contentArn:ContentSummary', contentSummary_contentArn - The Amazon Resource Name (ARN) of the content.

$sel:contentId:ContentSummary', contentSummary_contentId - The identifier of the content.

$sel:contentType:ContentSummary', contentSummary_contentType - The media type of the content.

$sel:knowledgeBaseArn:ContentSummary', contentSummary_knowledgeBaseArn - The Amazon Resource Name (ARN) of the knowledge base.

$sel:knowledgeBaseId:ContentSummary', contentSummary_knowledgeBaseId - The the identifier of the knowledge base.

$sel:metadata:ContentSummary', contentSummary_metadata - A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

$sel:name:ContentSummary', contentSummary_name - The name of the content.

$sel:revisionId:ContentSummary', contentSummary_revisionId - The identifier of the revision of the content.

$sel:status:ContentSummary', contentSummary_status - The status of the content.

$sel:title:ContentSummary', contentSummary_title - The title of the content.

contentSummary_tags :: Lens' ContentSummary (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

contentSummary_contentArn :: Lens' ContentSummary Text

The Amazon Resource Name (ARN) of the content.

contentSummary_contentId :: Lens' ContentSummary Text

The identifier of the content.

contentSummary_contentType :: Lens' ContentSummary Text

The media type of the content.

contentSummary_knowledgeBaseArn :: Lens' ContentSummary Text

The Amazon Resource Name (ARN) of the knowledge base.

contentSummary_knowledgeBaseId :: Lens' ContentSummary Text

The the identifier of the knowledge base.

contentSummary_metadata :: Lens' ContentSummary (HashMap Text Text)

A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

contentSummary_name :: Lens' ContentSummary Text

The name of the content.

contentSummary_revisionId :: Lens' ContentSummary Text

The identifier of the revision of the content.

contentSummary_status :: Lens' ContentSummary ContentStatus

The status of the content.

contentSummary_title :: Lens' ContentSummary Text

The title of the content.