Network.AWS.Wisdom.Types.KnowledgeBaseData

Description

Documentation

data KnowledgeBaseData

Information about the knowledge base.

See: newKnowledgeBaseData smart constructor.

Constructors

KnowledgeBaseData'

Fields

renderingConfiguration :: Maybe RenderingConfiguration

Information about how to render the content.

sourceConfiguration :: Maybe SourceConfiguration

Source configuration information about the knowledge base.

lastContentModificationTime :: Maybe POSIX

An epoch timestamp indicating the most recent content modification inside the knowledge base. If no content exists in a knowledge base, this value is unset.

description :: Maybe Text

The description.

serverSideEncryptionConfiguration :: Maybe ServerSideEncryptionConfiguration

The KMS key used for encryption.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

knowledgeBaseArn :: Text

The Amazon Resource Name (ARN) of the knowledge base.

knowledgeBaseId :: Text

The the identifier of the knowledge base.

knowledgeBaseType :: KnowledgeBaseType

The type of knowledge base.

name :: Text

The name of the knowledge base.

status :: KnowledgeBaseStatus

The status of the knowledge base.

newKnowledgeBaseData

Arguments

:: Text
$sel:knowledgeBaseArn:KnowledgeBaseData'

-> Text
$sel:knowledgeBaseId:KnowledgeBaseData'

-> KnowledgeBaseType
$sel:knowledgeBaseType:KnowledgeBaseData'

-> Text
$sel:name:KnowledgeBaseData'

-> KnowledgeBaseStatus
$sel:status:KnowledgeBaseData'

-> KnowledgeBaseData

Create a value of KnowledgeBaseData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:renderingConfiguration:KnowledgeBaseData', knowledgeBaseData_renderingConfiguration - Information about how to render the content.

$sel:sourceConfiguration:KnowledgeBaseData', knowledgeBaseData_sourceConfiguration - Source configuration information about the knowledge base.

$sel:lastContentModificationTime:KnowledgeBaseData', knowledgeBaseData_lastContentModificationTime - An epoch timestamp indicating the most recent content modification inside the knowledge base. If no content exists in a knowledge base, this value is unset.

$sel:description:KnowledgeBaseData', knowledgeBaseData_description - The description.

$sel:serverSideEncryptionConfiguration:KnowledgeBaseData', knowledgeBaseData_serverSideEncryptionConfiguration - The KMS key used for encryption.

$sel:tags:KnowledgeBaseData', knowledgeBaseData_tags - The tags used to organize, track, or control access for this resource.

$sel:knowledgeBaseArn:KnowledgeBaseData', knowledgeBaseData_knowledgeBaseArn - The Amazon Resource Name (ARN) of the knowledge base.

$sel:knowledgeBaseId:KnowledgeBaseData', knowledgeBaseData_knowledgeBaseId - The the identifier of the knowledge base.

$sel:knowledgeBaseType:KnowledgeBaseData', knowledgeBaseData_knowledgeBaseType - The type of knowledge base.

$sel:name:KnowledgeBaseData', knowledgeBaseData_name - The name of the knowledge base.

$sel:status:KnowledgeBaseData', knowledgeBaseData_status - The status of the knowledge base.

knowledgeBaseData_renderingConfiguration :: Lens' KnowledgeBaseData (Maybe RenderingConfiguration)

Information about how to render the content.

knowledgeBaseData_sourceConfiguration :: Lens' KnowledgeBaseData (Maybe SourceConfiguration)

Source configuration information about the knowledge base.

knowledgeBaseData_lastContentModificationTime :: Lens' KnowledgeBaseData (Maybe UTCTime)

An epoch timestamp indicating the most recent content modification inside the knowledge base. If no content exists in a knowledge base, this value is unset.

knowledgeBaseData_description :: Lens' KnowledgeBaseData (Maybe Text)

The description.

knowledgeBaseData_serverSideEncryptionConfiguration :: Lens' KnowledgeBaseData (Maybe ServerSideEncryptionConfiguration)

The KMS key used for encryption.

knowledgeBaseData_tags :: Lens' KnowledgeBaseData (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

knowledgeBaseData_knowledgeBaseArn :: Lens' KnowledgeBaseData Text

The Amazon Resource Name (ARN) of the knowledge base.

knowledgeBaseData_knowledgeBaseId :: Lens' KnowledgeBaseData Text

The the identifier of the knowledge base.

knowledgeBaseData_knowledgeBaseType :: Lens' KnowledgeBaseData KnowledgeBaseType

The type of knowledge base.

knowledgeBaseData_name :: Lens' KnowledgeBaseData Text

The name of the knowledge base.

knowledgeBaseData_status :: Lens' KnowledgeBaseData KnowledgeBaseStatus

The status of the knowledge base.