Network.AWS.Wisdom.Types.RenderingConfiguration

Description

Documentation

data RenderingConfiguration

Information about how to render the content.

See: newRenderingConfiguration smart constructor.

Constructors

RenderingConfiguration'

Fields

templateUri :: Maybe Text

A URI template containing exactly one variable in ${variableName} format. This can only be set for EXTERNAL knowledge bases. For Salesforce and ServiceNow, the variable must be one of the following:

Salesforce: Id, ArticleNumber, VersionNumber, Title, PublishStatus, or IsDeleted

ServiceNow: number, short_description, sys_mod_count, workflow_state, or active

 <p>The variable is replaced with the actual value for a piece of content when calling <a href="https://docs.aws.amazon.com/wisdom/latest/APIReference/API_GetContent.html">GetContent</a>. </p>

newRenderingConfiguration :: RenderingConfiguration

Create a value of RenderingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateUri:RenderingConfiguration', renderingConfiguration_templateUri - A URI template containing exactly one variable in ${variableName} format. This can only be set for EXTERNAL knowledge bases. For Salesforce and ServiceNow, the variable must be one of the following:

Salesforce: Id, ArticleNumber, VersionNumber, Title, PublishStatus, or IsDeleted

ServiceNow: number, short_description, sys_mod_count, workflow_state, or active

 <p>The variable is replaced with the actual value for a piece of content when calling <a href="https://docs.aws.amazon.com/wisdom/latest/APIReference/API_GetContent.html">GetContent</a>. </p>

renderingConfiguration_templateUri :: Lens' RenderingConfiguration (Maybe Text)

A URI template containing exactly one variable in ${variableName} format. This can only be set for EXTERNAL knowledge bases. For Salesforce and ServiceNow, the variable must be one of the following:

Salesforce: Id, ArticleNumber, VersionNumber, Title, PublishStatus, or IsDeleted

ServiceNow: number, short_description, sys_mod_count, workflow_state, or active

 <p>The variable is replaced with the actual value for a piece of content when calling <a href="https://docs.aws.amazon.com/wisdom/latest/APIReference/API_GetContent.html">GetContent</a>. </p>