Network.AWS.WorkDocs.ActivateUser

Description

Activates the specified user. Only active users can access Amazon WorkDocs.

Creating a Request

data ActivateUser

See: newActivateUser smart constructor.

Constructors

ActivateUser'

Fields

authenticationToken :: Maybe (Sensitive Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

userId :: Text

The ID of the user.

newActivateUser

Arguments

:: Text
$sel:userId:ActivateUser'

-> ActivateUser

Create a value of ActivateUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authenticationToken:ActivateUser', activateUser_authenticationToken - Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

$sel:userId:ActivateUser', activateUser_userId - The ID of the user.

Request Lenses

activateUser_authenticationToken :: Lens' ActivateUser (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

activateUser_userId :: Lens' ActivateUser Text

The ID of the user.

Destructuring the Response

data ActivateUserResponse

See: newActivateUserResponse smart constructor.

Constructors

ActivateUserResponse'

Fields

user :: Maybe User

The user information.

httpStatus :: Int

The response's http status code.

newActivateUserResponse

Arguments

:: Int
$sel:httpStatus:ActivateUserResponse'

-> ActivateUserResponse

Create a value of ActivateUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:user:ActivateUserResponse', activateUserResponse_user - The user information.

$sel:httpStatus:ActivateUserResponse', activateUserResponse_httpStatus - The response's http status code.

Response Lenses

activateUserResponse_user :: Lens' ActivateUserResponse (Maybe User)

The user information.

activateUserResponse_httpStatus :: Lens' ActivateUserResponse Int

The response's http status code.