Network.AWS.WorkDocs.AddResourcePermissions

Description

Creates a set of permissions for the specified folder or document. The resource permissions are overwritten if the principals already have different permissions.

Creating a Request

data AddResourcePermissions

See: newAddResourcePermissions smart constructor.

Constructors

AddResourcePermissions'

Fields

notificationOptions :: Maybe NotificationOptions

The notification options.

authenticationToken :: Maybe (Sensitive Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

resourceId :: Text

The ID of the resource.

principals :: [SharePrincipal]

The users, groups, or organization being granted permission.

newAddResourcePermissions

Arguments

:: Text
$sel:resourceId:AddResourcePermissions'

-> AddResourcePermissions

Create a value of AddResourcePermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notificationOptions:AddResourcePermissions', addResourcePermissions_notificationOptions - The notification options.

$sel:authenticationToken:AddResourcePermissions', addResourcePermissions_authenticationToken - Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

$sel:resourceId:AddResourcePermissions', addResourcePermissions_resourceId - The ID of the resource.

$sel:principals:AddResourcePermissions', addResourcePermissions_principals - The users, groups, or organization being granted permission.

Request Lenses

addResourcePermissions_notificationOptions :: Lens' AddResourcePermissions (Maybe NotificationOptions)

The notification options.

addResourcePermissions_authenticationToken :: Lens' AddResourcePermissions (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

addResourcePermissions_resourceId :: Lens' AddResourcePermissions Text

The ID of the resource.

addResourcePermissions_principals :: Lens' AddResourcePermissions [SharePrincipal]

The users, groups, or organization being granted permission.

Destructuring the Response

data AddResourcePermissionsResponse

See: newAddResourcePermissionsResponse smart constructor.

Constructors

AddResourcePermissionsResponse'

Fields

shareResults :: Maybe [ShareResult]

The share results.

httpStatus :: Int

The response's http status code.

newAddResourcePermissionsResponse

Arguments

:: Int
$sel:httpStatus:AddResourcePermissionsResponse'

-> AddResourcePermissionsResponse

Create a value of AddResourcePermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shareResults:AddResourcePermissionsResponse', addResourcePermissionsResponse_shareResults - The share results.

$sel:httpStatus:AddResourcePermissionsResponse', addResourcePermissionsResponse_httpStatus - The response's http status code.

Response Lenses

addResourcePermissionsResponse_shareResults :: Lens' AddResourcePermissionsResponse (Maybe [ShareResult])

The share results.

addResourcePermissionsResponse_httpStatus :: Lens' AddResourcePermissionsResponse Int

The response's http status code.