Network.AWS.WorkDocs.CreateComment

Description

Adds a new comment to the specified document version.

Creating a Request

data CreateComment

See: newCreateComment smart constructor.

Constructors

CreateComment'

Fields

parentId :: Maybe Text

The ID of the parent comment.

visibility :: Maybe CommentVisibilityType

The visibility of the comment. Options are either PRIVATE, where the comment is visible only to the comment author and document owner and co-owners, or PUBLIC, where the comment is visible to document owners, co-owners, and contributors.

threadId :: Maybe Text

The ID of the root comment in the thread.

authenticationToken :: Maybe (Sensitive Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

notifyCollaborators :: Maybe Bool

Set this parameter to TRUE to send an email out to the document collaborators after the comment is created.

documentId :: Text

The ID of the document.

versionId :: Text

The ID of the document version.

text :: Sensitive Text

The text of the comment.

newCreateComment

Arguments

:: Text
$sel:documentId:CreateComment'

-> Text
$sel:versionId:CreateComment'

-> Text
$sel:text:CreateComment'

-> CreateComment

Create a value of CreateComment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parentId:CreateComment', createComment_parentId - The ID of the parent comment.

$sel:visibility:CreateComment', createComment_visibility - The visibility of the comment. Options are either PRIVATE, where the comment is visible only to the comment author and document owner and co-owners, or PUBLIC, where the comment is visible to document owners, co-owners, and contributors.

$sel:threadId:CreateComment', createComment_threadId - The ID of the root comment in the thread.

$sel:authenticationToken:CreateComment', createComment_authenticationToken - Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

$sel:notifyCollaborators:CreateComment', createComment_notifyCollaborators - Set this parameter to TRUE to send an email out to the document collaborators after the comment is created.

$sel:documentId:CreateComment', createComment_documentId - The ID of the document.

$sel:versionId:CreateComment', createComment_versionId - The ID of the document version.

$sel:text:CreateComment', createComment_text - The text of the comment.

Request Lenses

createComment_parentId :: Lens' CreateComment (Maybe Text)

The ID of the parent comment.

createComment_visibility :: Lens' CreateComment (Maybe CommentVisibilityType)

The visibility of the comment. Options are either PRIVATE, where the comment is visible only to the comment author and document owner and co-owners, or PUBLIC, where the comment is visible to document owners, co-owners, and contributors.

createComment_threadId :: Lens' CreateComment (Maybe Text)

The ID of the root comment in the thread.

createComment_authenticationToken :: Lens' CreateComment (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

createComment_notifyCollaborators :: Lens' CreateComment (Maybe Bool)

Set this parameter to TRUE to send an email out to the document collaborators after the comment is created.

createComment_documentId :: Lens' CreateComment Text

The ID of the document.

createComment_versionId :: Lens' CreateComment Text

The ID of the document version.

createComment_text :: Lens' CreateComment Text

The text of the comment.

Destructuring the Response

data CreateCommentResponse

See: newCreateCommentResponse smart constructor.

Constructors

CreateCommentResponse'

Fields

comment :: Maybe Comment

The comment that has been created.

httpStatus :: Int

The response's http status code.

newCreateCommentResponse

Arguments

:: Int
$sel:httpStatus:CreateCommentResponse'

-> CreateCommentResponse

Create a value of CreateCommentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:comment:CreateCommentResponse', createCommentResponse_comment - The comment that has been created.

$sel:httpStatus:CreateCommentResponse', createCommentResponse_httpStatus - The response's http status code.

Response Lenses

createCommentResponse_comment :: Lens' CreateCommentResponse (Maybe Comment)

The comment that has been created.

createCommentResponse_httpStatus :: Lens' CreateCommentResponse Int

The response's http status code.