Network.AWS.WorkDocs.CreateNotificationSubscription

Description

Configure Amazon WorkDocs to use Amazon SNS notifications. The endpoint receives a confirmation message, and must confirm the subscription.

For more information, see Subscribe to Notifications in the Amazon WorkDocs Developer Guide.

Creating a Request

data CreateNotificationSubscription

See: newCreateNotificationSubscription smart constructor.

Constructors

CreateNotificationSubscription'

Fields

organizationId :: Text

The ID of the organization.

endpoint :: Text

The endpoint to receive the notifications. If the protocol is HTTPS, the endpoint is a URL that begins with https.

protocol :: SubscriptionProtocolType

The protocol to use. The supported value is https, which delivers JSON-encoded messages using HTTPS POST.

subscriptionType :: SubscriptionType

The notification type.

newCreateNotificationSubscription

Arguments

:: Text
$sel:organizationId:CreateNotificationSubscription'

-> Text
$sel:endpoint:CreateNotificationSubscription'

-> SubscriptionProtocolType
$sel:protocol:CreateNotificationSubscription'

-> SubscriptionType
$sel:subscriptionType:CreateNotificationSubscription'

-> CreateNotificationSubscription

Create a value of CreateNotificationSubscription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:organizationId:CreateNotificationSubscription', createNotificationSubscription_organizationId - The ID of the organization.

$sel:endpoint:CreateNotificationSubscription', createNotificationSubscription_endpoint - The endpoint to receive the notifications. If the protocol is HTTPS, the endpoint is a URL that begins with https.

$sel:protocol:CreateNotificationSubscription', createNotificationSubscription_protocol - The protocol to use. The supported value is https, which delivers JSON-encoded messages using HTTPS POST.

$sel:subscriptionType:CreateNotificationSubscription', createNotificationSubscription_subscriptionType - The notification type.

Request Lenses

createNotificationSubscription_organizationId :: Lens' CreateNotificationSubscription Text

The ID of the organization.

createNotificationSubscription_endpoint :: Lens' CreateNotificationSubscription Text

The endpoint to receive the notifications. If the protocol is HTTPS, the endpoint is a URL that begins with https.

createNotificationSubscription_protocol :: Lens' CreateNotificationSubscription SubscriptionProtocolType

The protocol to use. The supported value is https, which delivers JSON-encoded messages using HTTPS POST.

createNotificationSubscription_subscriptionType :: Lens' CreateNotificationSubscription SubscriptionType

The notification type.

Destructuring the Response

data CreateNotificationSubscriptionResponse

See: newCreateNotificationSubscriptionResponse smart constructor.

Constructors

CreateNotificationSubscriptionResponse'

Fields

subscription :: Maybe Subscription

The subscription.

httpStatus :: Int

The response's http status code.

newCreateNotificationSubscriptionResponse

Arguments

:: Int
$sel:httpStatus:CreateNotificationSubscriptionResponse'

-> CreateNotificationSubscriptionResponse

Create a value of CreateNotificationSubscriptionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscription:CreateNotificationSubscriptionResponse', createNotificationSubscriptionResponse_subscription - The subscription.

$sel:httpStatus:CreateNotificationSubscriptionResponse', createNotificationSubscriptionResponse_httpStatus - The response's http status code.

Response Lenses

createNotificationSubscriptionResponse_subscription :: Lens' CreateNotificationSubscriptionResponse (Maybe Subscription)

The subscription.

createNotificationSubscriptionResponse_httpStatus :: Lens' CreateNotificationSubscriptionResponse Int

The response's http status code.