Network.AWS.WorkDocs.DeleteLabels

Description

Deletes the specified list of labels from a resource.

Creating a Request

data DeleteLabels

See: newDeleteLabels smart constructor.

Constructors

DeleteLabels'

Fields

deleteAll :: Maybe Bool

Flag to request removal of all labels from the specified resource.

authenticationToken :: Maybe (Sensitive Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

labels :: Maybe [Text]

List of labels to delete from the resource.

resourceId :: Text

The ID of the resource.

newDeleteLabels

Arguments

:: Text
$sel:resourceId:DeleteLabels'

-> DeleteLabels

Create a value of DeleteLabels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deleteAll:DeleteLabels', deleteLabels_deleteAll - Flag to request removal of all labels from the specified resource.

$sel:authenticationToken:DeleteLabels', deleteLabels_authenticationToken - Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

$sel:labels:DeleteLabels', deleteLabels_labels - List of labels to delete from the resource.

$sel:resourceId:DeleteLabels', deleteLabels_resourceId - The ID of the resource.

Request Lenses

deleteLabels_deleteAll :: Lens' DeleteLabels (Maybe Bool)

Flag to request removal of all labels from the specified resource.

deleteLabels_authenticationToken :: Lens' DeleteLabels (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

deleteLabels_labels :: Lens' DeleteLabels (Maybe [Text])

List of labels to delete from the resource.

deleteLabels_resourceId :: Lens' DeleteLabels Text

The ID of the resource.

Destructuring the Response

data DeleteLabelsResponse

See: newDeleteLabelsResponse smart constructor.

Constructors

DeleteLabelsResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteLabelsResponse

Arguments

:: Int
$sel:httpStatus:DeleteLabelsResponse'

-> DeleteLabelsResponse

Create a value of DeleteLabelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteLabelsResponse', deleteLabelsResponse_httpStatus - The response's http status code.

Response Lenses

deleteLabelsResponse_httpStatus :: Lens' DeleteLabelsResponse Int

The response's http status code.