Network.AWS.WorkDocs.DescribeActivities

Description

Describes the user activities in a specified time period.

This operation returns paginated results.

Creating a Request

data DescribeActivities

See: newDescribeActivities smart constructor.

Constructors

DescribeActivities'

Fields

resourceId :: Maybe Text

The document or folder ID for which to describe activity types.

organizationId :: Maybe Text

The ID of the organization. This is a mandatory parameter when using administrative API (SigV4) requests.

includeIndirectActivities :: Maybe Bool

Includes indirect activities. An indirect activity results from a direct activity performed on a parent resource. For example, sharing a parent folder (the direct activity) shares all of the subfolders and documents within the parent folder (the indirect activity).

startTime :: Maybe POSIX

The timestamp that determines the starting time of the activities. The response includes the activities performed after the specified timestamp.

endTime :: Maybe POSIX

The timestamp that determines the end time of the activities. The response includes the activities performed before the specified timestamp.

userId :: Maybe Text

The ID of the user who performed the action. The response includes activities pertaining to this user. This is an optional parameter and is only applicable for administrative API (SigV4) requests.

activityTypes :: Maybe Text

Specifies which activity types to include in the response. If this field is left empty, all activity types are returned.

authenticationToken :: Maybe (Sensitive Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

limit :: Maybe Natural

The maximum number of items to return.

marker :: Maybe Text

The marker for the next set of results.

newDescribeActivities :: DescribeActivities

Create a value of DescribeActivities with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:DescribeActivities', describeActivities_resourceId - The document or folder ID for which to describe activity types.

$sel:organizationId:DescribeActivities', describeActivities_organizationId - The ID of the organization. This is a mandatory parameter when using administrative API (SigV4) requests.

$sel:includeIndirectActivities:DescribeActivities', describeActivities_includeIndirectActivities - Includes indirect activities. An indirect activity results from a direct activity performed on a parent resource. For example, sharing a parent folder (the direct activity) shares all of the subfolders and documents within the parent folder (the indirect activity).

$sel:startTime:DescribeActivities', describeActivities_startTime - The timestamp that determines the starting time of the activities. The response includes the activities performed after the specified timestamp.

$sel:endTime:DescribeActivities', describeActivities_endTime - The timestamp that determines the end time of the activities. The response includes the activities performed before the specified timestamp.

$sel:userId:DescribeActivities', describeActivities_userId - The ID of the user who performed the action. The response includes activities pertaining to this user. This is an optional parameter and is only applicable for administrative API (SigV4) requests.

$sel:activityTypes:DescribeActivities', describeActivities_activityTypes - Specifies which activity types to include in the response. If this field is left empty, all activity types are returned.

$sel:authenticationToken:DescribeActivities', describeActivities_authenticationToken - Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

$sel:limit:DescribeActivities', describeActivities_limit - The maximum number of items to return.

$sel:marker:DescribeActivities', describeActivities_marker - The marker for the next set of results.

Request Lenses

describeActivities_resourceId :: Lens' DescribeActivities (Maybe Text)

The document or folder ID for which to describe activity types.

describeActivities_organizationId :: Lens' DescribeActivities (Maybe Text)

The ID of the organization. This is a mandatory parameter when using administrative API (SigV4) requests.

describeActivities_includeIndirectActivities :: Lens' DescribeActivities (Maybe Bool)

Includes indirect activities. An indirect activity results from a direct activity performed on a parent resource. For example, sharing a parent folder (the direct activity) shares all of the subfolders and documents within the parent folder (the indirect activity).

describeActivities_startTime :: Lens' DescribeActivities (Maybe UTCTime)

The timestamp that determines the starting time of the activities. The response includes the activities performed after the specified timestamp.

describeActivities_endTime :: Lens' DescribeActivities (Maybe UTCTime)

The timestamp that determines the end time of the activities. The response includes the activities performed before the specified timestamp.

describeActivities_userId :: Lens' DescribeActivities (Maybe Text)

The ID of the user who performed the action. The response includes activities pertaining to this user. This is an optional parameter and is only applicable for administrative API (SigV4) requests.

describeActivities_activityTypes :: Lens' DescribeActivities (Maybe Text)

Specifies which activity types to include in the response. If this field is left empty, all activity types are returned.

describeActivities_authenticationToken :: Lens' DescribeActivities (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

describeActivities_limit :: Lens' DescribeActivities (Maybe Natural)

The maximum number of items to return.

describeActivities_marker :: Lens' DescribeActivities (Maybe Text)

The marker for the next set of results.

Destructuring the Response

data DescribeActivitiesResponse

See: newDescribeActivitiesResponse smart constructor.

Constructors

DescribeActivitiesResponse'

Fields

userActivities :: Maybe [Activity]

The list of activities for the specified user and time period.

marker :: Maybe Text

The marker for the next set of results.

httpStatus :: Int

The response's http status code.

newDescribeActivitiesResponse

Arguments

:: Int
$sel:httpStatus:DescribeActivitiesResponse'

-> DescribeActivitiesResponse

Create a value of DescribeActivitiesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userActivities:DescribeActivitiesResponse', describeActivitiesResponse_userActivities - The list of activities for the specified user and time period.

$sel:marker:DescribeActivities', describeActivitiesResponse_marker - The marker for the next set of results.

$sel:httpStatus:DescribeActivitiesResponse', describeActivitiesResponse_httpStatus - The response's http status code.

Response Lenses

describeActivitiesResponse_userActivities :: Lens' DescribeActivitiesResponse (Maybe [Activity])

The list of activities for the specified user and time period.

describeActivitiesResponse_marker :: Lens' DescribeActivitiesResponse (Maybe Text)

The marker for the next set of results.

describeActivitiesResponse_httpStatus :: Lens' DescribeActivitiesResponse Int

The response's http status code.