Network.AWS.WorkDocs.DescribeDocumentVersions

Description

Retrieves the document versions for the specified document.

By default, only active versions are returned.

This operation returns paginated results.

Creating a Request

data DescribeDocumentVersions

See: newDescribeDocumentVersions smart constructor.

Constructors

DescribeDocumentVersions'

Fields

include :: Maybe Text

A comma-separated list of values. Specify "INITIALIZED" to include incomplete versions.

fields :: Maybe Text

Specify "SOURCE" to include initialized versions and a URL for the source document.

authenticationToken :: Maybe (Sensitive Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

limit :: Maybe Natural

The maximum number of versions to return with this call.

marker :: Maybe Text

The marker for the next set of results. (You received this marker from a previous call.)

documentId :: Text

The ID of the document.

newDescribeDocumentVersions

Arguments

:: Text
$sel:documentId:DescribeDocumentVersions'

-> DescribeDocumentVersions

Create a value of DescribeDocumentVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:include:DescribeDocumentVersions', describeDocumentVersions_include - A comma-separated list of values. Specify "INITIALIZED" to include incomplete versions.

$sel:fields:DescribeDocumentVersions', describeDocumentVersions_fields - Specify "SOURCE" to include initialized versions and a URL for the source document.

$sel:authenticationToken:DescribeDocumentVersions', describeDocumentVersions_authenticationToken - Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

$sel:limit:DescribeDocumentVersions', describeDocumentVersions_limit - The maximum number of versions to return with this call.

$sel:marker:DescribeDocumentVersions', describeDocumentVersions_marker - The marker for the next set of results. (You received this marker from a previous call.)

$sel:documentId:DescribeDocumentVersions', describeDocumentVersions_documentId - The ID of the document.

Request Lenses

describeDocumentVersions_include :: Lens' DescribeDocumentVersions (Maybe Text)

A comma-separated list of values. Specify "INITIALIZED" to include incomplete versions.

describeDocumentVersions_fields :: Lens' DescribeDocumentVersions (Maybe Text)

Specify "SOURCE" to include initialized versions and a URL for the source document.

describeDocumentVersions_authenticationToken :: Lens' DescribeDocumentVersions (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

describeDocumentVersions_limit :: Lens' DescribeDocumentVersions (Maybe Natural)

The maximum number of versions to return with this call.

describeDocumentVersions_marker :: Lens' DescribeDocumentVersions (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

describeDocumentVersions_documentId :: Lens' DescribeDocumentVersions Text

The ID of the document.

Destructuring the Response

data DescribeDocumentVersionsResponse

See: newDescribeDocumentVersionsResponse smart constructor.

Constructors

DescribeDocumentVersionsResponse'

Fields

documentVersions :: Maybe [DocumentVersionMetadata]

The document versions.

marker :: Maybe Text

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

httpStatus :: Int

The response's http status code.

newDescribeDocumentVersionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeDocumentVersionsResponse'

-> DescribeDocumentVersionsResponse

Create a value of DescribeDocumentVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:documentVersions:DescribeDocumentVersionsResponse', describeDocumentVersionsResponse_documentVersions - The document versions.

$sel:marker:DescribeDocumentVersions', describeDocumentVersionsResponse_marker - The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

$sel:httpStatus:DescribeDocumentVersionsResponse', describeDocumentVersionsResponse_httpStatus - The response's http status code.

Response Lenses

describeDocumentVersionsResponse_documentVersions :: Lens' DescribeDocumentVersionsResponse (Maybe [DocumentVersionMetadata])

The document versions.

describeDocumentVersionsResponse_marker :: Lens' DescribeDocumentVersionsResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

describeDocumentVersionsResponse_httpStatus :: Lens' DescribeDocumentVersionsResponse Int

The response's http status code.