Network.AWS.WorkDocs.DescribeFolderContents

Description

Describes the contents of the specified folder, including its documents and subfolders.

By default, Amazon WorkDocs returns the first 100 active document and folder metadata items. If there are more results, the response includes a marker that you can use to request the next set of results. You can also request initialized documents.

This operation returns paginated results.

Creating a Request

data DescribeFolderContents

See: newDescribeFolderContents smart constructor.

Constructors

DescribeFolderContents'

Fields

include :: Maybe Text

The contents to include. Specify "INITIALIZED" to include initialized documents.

order :: Maybe OrderType

The order for the contents of the folder.

authenticationToken :: Maybe (Sensitive Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

type' :: Maybe FolderContentType

The type of items.

limit :: Maybe Natural

The maximum number of items to return with this call.

sort :: Maybe ResourceSortType

The sorting criteria.

marker :: Maybe Text

The marker for the next set of results. This marker was received from a previous call.

folderId :: Text

The ID of the folder.

newDescribeFolderContents

Arguments

:: Text
$sel:folderId:DescribeFolderContents'

-> DescribeFolderContents

Create a value of DescribeFolderContents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:include:DescribeFolderContents', describeFolderContents_include - The contents to include. Specify "INITIALIZED" to include initialized documents.

$sel:order:DescribeFolderContents', describeFolderContents_order - The order for the contents of the folder.

$sel:authenticationToken:DescribeFolderContents', describeFolderContents_authenticationToken - Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

$sel:type':DescribeFolderContents', describeFolderContents_type - The type of items.

$sel:limit:DescribeFolderContents', describeFolderContents_limit - The maximum number of items to return with this call.

$sel:sort:DescribeFolderContents', describeFolderContents_sort - The sorting criteria.

$sel:marker:DescribeFolderContents', describeFolderContents_marker - The marker for the next set of results. This marker was received from a previous call.

$sel:folderId:DescribeFolderContents', describeFolderContents_folderId - The ID of the folder.

Request Lenses

describeFolderContents_include :: Lens' DescribeFolderContents (Maybe Text)

The contents to include. Specify "INITIALIZED" to include initialized documents.

describeFolderContents_order :: Lens' DescribeFolderContents (Maybe OrderType)

The order for the contents of the folder.

describeFolderContents_authenticationToken :: Lens' DescribeFolderContents (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

describeFolderContents_type :: Lens' DescribeFolderContents (Maybe FolderContentType)

The type of items.

describeFolderContents_limit :: Lens' DescribeFolderContents (Maybe Natural)

The maximum number of items to return with this call.

describeFolderContents_sort :: Lens' DescribeFolderContents (Maybe ResourceSortType)

The sorting criteria.

describeFolderContents_marker :: Lens' DescribeFolderContents (Maybe Text)

The marker for the next set of results. This marker was received from a previous call.

describeFolderContents_folderId :: Lens' DescribeFolderContents Text

The ID of the folder.

Destructuring the Response

data DescribeFolderContentsResponse

See: newDescribeFolderContentsResponse smart constructor.

Constructors

DescribeFolderContentsResponse'

Fields

documents :: Maybe [DocumentMetadata]

The documents in the specified folder.

folders :: Maybe [FolderMetadata]

The subfolders in the specified folder.

marker :: Maybe Text

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

httpStatus :: Int

The response's http status code.

newDescribeFolderContentsResponse

Arguments

:: Int
$sel:httpStatus:DescribeFolderContentsResponse'

-> DescribeFolderContentsResponse

Create a value of DescribeFolderContentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:documents:DescribeFolderContentsResponse', describeFolderContentsResponse_documents - The documents in the specified folder.

$sel:folders:DescribeFolderContentsResponse', describeFolderContentsResponse_folders - The subfolders in the specified folder.

$sel:marker:DescribeFolderContents', describeFolderContentsResponse_marker - The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

$sel:httpStatus:DescribeFolderContentsResponse', describeFolderContentsResponse_httpStatus - The response's http status code.

Response Lenses

describeFolderContentsResponse_documents :: Lens' DescribeFolderContentsResponse (Maybe [DocumentMetadata])

The documents in the specified folder.

describeFolderContentsResponse_folders :: Lens' DescribeFolderContentsResponse (Maybe [FolderMetadata])

The subfolders in the specified folder.

describeFolderContentsResponse_marker :: Lens' DescribeFolderContentsResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

describeFolderContentsResponse_httpStatus :: Lens' DescribeFolderContentsResponse Int

The response's http status code.