Network.AWS.WorkDocs.DescribeGroups

Description

Describes the groups specified by the query. Groups are defined by the underlying Active Directory.

This operation returns paginated results.

Creating a Request

data DescribeGroups

See: newDescribeGroups smart constructor.

Constructors

DescribeGroups'

Fields

authenticationToken :: Maybe (Sensitive Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

marker :: Maybe Text

The marker for the next set of results. (You received this marker from a previous call.)

limit :: Maybe Natural

The maximum number of items to return with this call.

organizationId :: Maybe Text

The ID of the organization.

searchQuery :: Sensitive Text

A query to describe groups by group name.

newDescribeGroups

Arguments

:: Text
$sel:searchQuery:DescribeGroups'

-> DescribeGroups

Create a value of DescribeGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authenticationToken:DescribeGroups', describeGroups_authenticationToken - Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

$sel:marker:DescribeGroups', describeGroups_marker - The marker for the next set of results. (You received this marker from a previous call.)

$sel:limit:DescribeGroups', describeGroups_limit - The maximum number of items to return with this call.

$sel:organizationId:DescribeGroups', describeGroups_organizationId - The ID of the organization.

$sel:searchQuery:DescribeGroups', describeGroups_searchQuery - A query to describe groups by group name.

Request Lenses

describeGroups_authenticationToken :: Lens' DescribeGroups (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

describeGroups_marker :: Lens' DescribeGroups (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

describeGroups_limit :: Lens' DescribeGroups (Maybe Natural)

The maximum number of items to return with this call.

describeGroups_organizationId :: Lens' DescribeGroups (Maybe Text)

The ID of the organization.

describeGroups_searchQuery :: Lens' DescribeGroups Text

A query to describe groups by group name.

Destructuring the Response

data DescribeGroupsResponse

See: newDescribeGroupsResponse smart constructor.

Constructors

DescribeGroupsResponse'

Fields

groups :: Maybe [GroupMetadata]

The list of groups.

marker :: Maybe Text

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

httpStatus :: Int

The response's http status code.

newDescribeGroupsResponse

Arguments

:: Int
$sel:httpStatus:DescribeGroupsResponse'

-> DescribeGroupsResponse

Create a value of DescribeGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groups:DescribeGroupsResponse', describeGroupsResponse_groups - The list of groups.

$sel:marker:DescribeGroups', describeGroupsResponse_marker - The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

$sel:httpStatus:DescribeGroupsResponse', describeGroupsResponse_httpStatus - The response's http status code.

Response Lenses

describeGroupsResponse_groups :: Lens' DescribeGroupsResponse (Maybe [GroupMetadata])

The list of groups.

describeGroupsResponse_marker :: Lens' DescribeGroupsResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

describeGroupsResponse_httpStatus :: Lens' DescribeGroupsResponse Int

The response's http status code.