Network.AWS.WorkDocs.DescribeNotificationSubscriptions

Description

Lists the specified notification subscriptions.

This operation returns paginated results.

Creating a Request

data DescribeNotificationSubscriptions

See: newDescribeNotificationSubscriptions smart constructor.

Constructors

DescribeNotificationSubscriptions'

Fields

limit :: Maybe Natural

The maximum number of items to return with this call.

marker :: Maybe Text

The marker for the next set of results. (You received this marker from a previous call.)

organizationId :: Text

The ID of the organization.

newDescribeNotificationSubscriptions

Arguments

:: Text
$sel:organizationId:DescribeNotificationSubscriptions'

-> DescribeNotificationSubscriptions

Create a value of DescribeNotificationSubscriptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:limit:DescribeNotificationSubscriptions', describeNotificationSubscriptions_limit - The maximum number of items to return with this call.

$sel:marker:DescribeNotificationSubscriptions', describeNotificationSubscriptions_marker - The marker for the next set of results. (You received this marker from a previous call.)

$sel:organizationId:DescribeNotificationSubscriptions', describeNotificationSubscriptions_organizationId - The ID of the organization.

Request Lenses

describeNotificationSubscriptions_limit :: Lens' DescribeNotificationSubscriptions (Maybe Natural)

The maximum number of items to return with this call.

describeNotificationSubscriptions_marker :: Lens' DescribeNotificationSubscriptions (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

describeNotificationSubscriptions_organizationId :: Lens' DescribeNotificationSubscriptions Text

The ID of the organization.

Destructuring the Response

data DescribeNotificationSubscriptionsResponse

See: newDescribeNotificationSubscriptionsResponse smart constructor.

Constructors

DescribeNotificationSubscriptionsResponse'

Fields

subscriptions :: Maybe [Subscription]

The subscriptions.

marker :: Maybe Text

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

httpStatus :: Int

The response's http status code.

newDescribeNotificationSubscriptionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeNotificationSubscriptionsResponse'

-> DescribeNotificationSubscriptionsResponse

Create a value of DescribeNotificationSubscriptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscriptions:DescribeNotificationSubscriptionsResponse', describeNotificationSubscriptionsResponse_subscriptions - The subscriptions.

$sel:marker:DescribeNotificationSubscriptions', describeNotificationSubscriptionsResponse_marker - The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

$sel:httpStatus:DescribeNotificationSubscriptionsResponse', describeNotificationSubscriptionsResponse_httpStatus - The response's http status code.

Response Lenses

describeNotificationSubscriptionsResponse_subscriptions :: Lens' DescribeNotificationSubscriptionsResponse (Maybe [Subscription])

The subscriptions.

describeNotificationSubscriptionsResponse_marker :: Lens' DescribeNotificationSubscriptionsResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

describeNotificationSubscriptionsResponse_httpStatus :: Lens' DescribeNotificationSubscriptionsResponse Int

The response's http status code.