Network.AWS.WorkDocs.DescribeUsers

Description

Describes the specified users. You can describe all users or filter the results (for example, by status or organization).

By default, Amazon WorkDocs returns the first 24 active or pending users. If there are more results, the response includes a marker that you can use to request the next set of results.

This operation returns paginated results.

Creating a Request

data DescribeUsers

See: newDescribeUsers smart constructor.

Constructors

DescribeUsers'

Fields

include :: Maybe UserFilterType

The state of the users. Specify "ALL" to include inactive users.

userIds :: Maybe Text

The IDs of the users.

authenticationToken :: Maybe (Sensitive Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

sort :: Maybe UserSortType

The sorting criteria.

marker :: Maybe Text

The marker for the next set of results. (You received this marker from a previous call.)

query :: Maybe (Sensitive Text)

A query to filter users by user name.

limit :: Maybe Natural

The maximum number of items to return.

order :: Maybe OrderType

The order for the results.

organizationId :: Maybe Text

The ID of the organization.

fields :: Maybe Text

A comma-separated list of values. Specify "STORAGE_METADATA" to include the user storage quota and utilization information.

newDescribeUsers :: DescribeUsers

Create a value of DescribeUsers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:include:DescribeUsers', describeUsers_include - The state of the users. Specify "ALL" to include inactive users.

$sel:userIds:DescribeUsers', describeUsers_userIds - The IDs of the users.

$sel:authenticationToken:DescribeUsers', describeUsers_authenticationToken - Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

$sel:sort:DescribeUsers', describeUsers_sort - The sorting criteria.

$sel:marker:DescribeUsers', describeUsers_marker - The marker for the next set of results. (You received this marker from a previous call.)

$sel:query:DescribeUsers', describeUsers_query - A query to filter users by user name.

$sel:limit:DescribeUsers', describeUsers_limit - The maximum number of items to return.

$sel:order:DescribeUsers', describeUsers_order - The order for the results.

$sel:organizationId:DescribeUsers', describeUsers_organizationId - The ID of the organization.

$sel:fields:DescribeUsers', describeUsers_fields - A comma-separated list of values. Specify "STORAGE_METADATA" to include the user storage quota and utilization information.

Request Lenses

describeUsers_include :: Lens' DescribeUsers (Maybe UserFilterType)

The state of the users. Specify "ALL" to include inactive users.

describeUsers_userIds :: Lens' DescribeUsers (Maybe Text)

The IDs of the users.

describeUsers_authenticationToken :: Lens' DescribeUsers (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

describeUsers_sort :: Lens' DescribeUsers (Maybe UserSortType)

The sorting criteria.

describeUsers_marker :: Lens' DescribeUsers (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

describeUsers_query :: Lens' DescribeUsers (Maybe Text)

A query to filter users by user name.

describeUsers_limit :: Lens' DescribeUsers (Maybe Natural)

The maximum number of items to return.

describeUsers_order :: Lens' DescribeUsers (Maybe OrderType)

The order for the results.

describeUsers_organizationId :: Lens' DescribeUsers (Maybe Text)

The ID of the organization.

describeUsers_fields :: Lens' DescribeUsers (Maybe Text)

A comma-separated list of values. Specify "STORAGE_METADATA" to include the user storage quota and utilization information.

Destructuring the Response

data DescribeUsersResponse

See: newDescribeUsersResponse smart constructor.

Constructors

DescribeUsersResponse'

Fields

users :: Maybe [User]

The users.

totalNumberOfUsers :: Maybe Integer

The total number of users included in the results.

marker :: Maybe Text

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

httpStatus :: Int

The response's http status code.

newDescribeUsersResponse

Arguments

:: Int
$sel:httpStatus:DescribeUsersResponse'

-> DescribeUsersResponse

Create a value of DescribeUsersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:users:DescribeUsersResponse', describeUsersResponse_users - The users.

$sel:totalNumberOfUsers:DescribeUsersResponse', describeUsersResponse_totalNumberOfUsers - The total number of users included in the results.

$sel:marker:DescribeUsers', describeUsersResponse_marker - The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

$sel:httpStatus:DescribeUsersResponse', describeUsersResponse_httpStatus - The response's http status code.

Response Lenses

describeUsersResponse_users :: Lens' DescribeUsersResponse (Maybe [User])

The users.

describeUsersResponse_totalNumberOfUsers :: Lens' DescribeUsersResponse (Maybe Integer)

The total number of users included in the results.

describeUsersResponse_marker :: Lens' DescribeUsersResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

describeUsersResponse_httpStatus :: Lens' DescribeUsersResponse Int

The response's http status code.