Network.AWS.WorkDocs.GetCurrentUser

Description

Retrieves details of the current user for whom the authentication token was generated. This is not a valid action for SigV4 (administrative API) clients.

This action requires an authentication token. To get an authentication token, register an application with Amazon WorkDocs. For more information, see Authentication and Access Control for User Applications in the Amazon WorkDocs Developer Guide.

Creating a Request

data GetCurrentUser

See: newGetCurrentUser smart constructor.

Constructors

GetCurrentUser'

Fields

authenticationToken :: Sensitive Text

Amazon WorkDocs authentication token.

newGetCurrentUser

Arguments

:: Text
$sel:authenticationToken:GetCurrentUser'

-> GetCurrentUser

Create a value of GetCurrentUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authenticationToken:GetCurrentUser', getCurrentUser_authenticationToken - Amazon WorkDocs authentication token.

Request Lenses

getCurrentUser_authenticationToken :: Lens' GetCurrentUser Text

Amazon WorkDocs authentication token.

Destructuring the Response

data GetCurrentUserResponse

See: newGetCurrentUserResponse smart constructor.

Constructors

GetCurrentUserResponse'

Fields

user :: Maybe User

Metadata of the user.

httpStatus :: Int

The response's http status code.

newGetCurrentUserResponse

Arguments

:: Int
$sel:httpStatus:GetCurrentUserResponse'

-> GetCurrentUserResponse

Create a value of GetCurrentUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:user:GetCurrentUserResponse', getCurrentUserResponse_user - Metadata of the user.

$sel:httpStatus:GetCurrentUserResponse', getCurrentUserResponse_httpStatus - The response's http status code.

Response Lenses

getCurrentUserResponse_user :: Lens' GetCurrentUserResponse (Maybe User)

Metadata of the user.

getCurrentUserResponse_httpStatus :: Lens' GetCurrentUserResponse Int

The response's http status code.