Network.AWS.WorkDocs.GetDocumentPath

Description

Retrieves the path information (the hierarchy from the root folder) for the requested document.

By default, Amazon WorkDocs returns a maximum of 100 levels upwards from the requested document and only includes the IDs of the parent folders in the path. You can limit the maximum number of levels. You can also request the names of the parent folders.

Creating a Request

data GetDocumentPath

See: newGetDocumentPath smart constructor.

Constructors

GetDocumentPath'

Fields

authenticationToken :: Maybe (Sensitive Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

marker :: Maybe Text

This value is not supported.

limit :: Maybe Natural

The maximum number of levels in the hierarchy to return.

fields :: Maybe Text

A comma-separated list of values. Specify NAME to include the names of the parent folders.

documentId :: Text

The ID of the document.

newGetDocumentPath

Arguments

:: Text
$sel:documentId:GetDocumentPath'

-> GetDocumentPath

Create a value of GetDocumentPath with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authenticationToken:GetDocumentPath', getDocumentPath_authenticationToken - Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

$sel:marker:GetDocumentPath', getDocumentPath_marker - This value is not supported.

$sel:limit:GetDocumentPath', getDocumentPath_limit - The maximum number of levels in the hierarchy to return.

$sel:fields:GetDocumentPath', getDocumentPath_fields - A comma-separated list of values. Specify NAME to include the names of the parent folders.

$sel:documentId:GetDocumentPath', getDocumentPath_documentId - The ID of the document.

Request Lenses

getDocumentPath_authenticationToken :: Lens' GetDocumentPath (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

getDocumentPath_marker :: Lens' GetDocumentPath (Maybe Text)

This value is not supported.

getDocumentPath_limit :: Lens' GetDocumentPath (Maybe Natural)

The maximum number of levels in the hierarchy to return.

getDocumentPath_fields :: Lens' GetDocumentPath (Maybe Text)

A comma-separated list of values. Specify NAME to include the names of the parent folders.

getDocumentPath_documentId :: Lens' GetDocumentPath Text

The ID of the document.

Destructuring the Response

data GetDocumentPathResponse

See: newGetDocumentPathResponse smart constructor.

Constructors

GetDocumentPathResponse'

Fields

path :: Maybe ResourcePath

The path information.

httpStatus :: Int

The response's http status code.

newGetDocumentPathResponse

Arguments

:: Int
$sel:httpStatus:GetDocumentPathResponse'

-> GetDocumentPathResponse

Create a value of GetDocumentPathResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:path:GetDocumentPathResponse', getDocumentPathResponse_path - The path information.

$sel:httpStatus:GetDocumentPathResponse', getDocumentPathResponse_httpStatus - The response's http status code.

Response Lenses

getDocumentPathResponse_path :: Lens' GetDocumentPathResponse (Maybe ResourcePath)

The path information.

getDocumentPathResponse_httpStatus :: Lens' GetDocumentPathResponse Int

The response's http status code.