Network.AWS.WorkDocs.GetFolder

Description

Retrieves the metadata of the specified folder.

Creating a Request

data GetFolder

See: newGetFolder smart constructor.

Constructors

GetFolder'

Fields

includeCustomMetadata :: Maybe Bool

Set to TRUE to include custom metadata in the response.

authenticationToken :: Maybe (Sensitive Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

folderId :: Text

The ID of the folder.

newGetFolder

Arguments

:: Text
$sel:folderId:GetFolder'

-> GetFolder

Create a value of GetFolder with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeCustomMetadata:GetFolder', getFolder_includeCustomMetadata - Set to TRUE to include custom metadata in the response.

$sel:authenticationToken:GetFolder', getFolder_authenticationToken - Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

$sel:folderId:GetFolder', getFolder_folderId - The ID of the folder.

Request Lenses

getFolder_includeCustomMetadata :: Lens' GetFolder (Maybe Bool)

Set to TRUE to include custom metadata in the response.

getFolder_authenticationToken :: Lens' GetFolder (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

getFolder_folderId :: Lens' GetFolder Text

The ID of the folder.

Destructuring the Response

data GetFolderResponse

See: newGetFolderResponse smart constructor.

Constructors

GetFolderResponse'

Fields

metadata :: Maybe FolderMetadata

The metadata of the folder.

customMetadata :: Maybe (HashMap Text Text)

The custom metadata on the folder.

httpStatus :: Int

The response's http status code.

newGetFolderResponse

Arguments

:: Int
$sel:httpStatus:GetFolderResponse'

-> GetFolderResponse

Create a value of GetFolderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metadata:GetFolderResponse', getFolderResponse_metadata - The metadata of the folder.

$sel:customMetadata:GetFolderResponse', getFolderResponse_customMetadata - The custom metadata on the folder.

$sel:httpStatus:GetFolderResponse', getFolderResponse_httpStatus - The response's http status code.

Response Lenses

getFolderResponse_metadata :: Lens' GetFolderResponse (Maybe FolderMetadata)

The metadata of the folder.

getFolderResponse_customMetadata :: Lens' GetFolderResponse (Maybe (HashMap Text Text))

The custom metadata on the folder.

getFolderResponse_httpStatus :: Lens' GetFolderResponse Int

The response's http status code.