Network.AWS.WorkDocs.InitiateDocumentVersionUpload

Description

Creates a new document object and version object.

The client specifies the parent folder ID and name of the document to upload. The ID is optionally specified when creating a new version of an existing document. This is the first step to upload a document. Next, upload the document to the URL returned from the call, and then call UpdateDocumentVersion.

To cancel the document upload, call AbortDocumentVersionUpload.

Creating a Request

data InitiateDocumentVersionUpload

See: newInitiateDocumentVersionUpload smart constructor.

Constructors

InitiateDocumentVersionUpload'

Fields

contentType :: Maybe Text

The content type of the document.

contentModifiedTimestamp :: Maybe POSIX

The timestamp when the content of the document was modified.

id :: Maybe Text

The ID of the document.

name :: Maybe Text

The name of the document.

contentCreatedTimestamp :: Maybe POSIX

The timestamp when the content of the document was originally created.

documentSizeInBytes :: Maybe Integer

The size of the document, in bytes.

authenticationToken :: Maybe (Sensitive Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

parentFolderId :: Text

The ID of the parent folder.

newInitiateDocumentVersionUpload

Arguments

:: Text
$sel:parentFolderId:InitiateDocumentVersionUpload'

-> InitiateDocumentVersionUpload

Create a value of InitiateDocumentVersionUpload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentType:InitiateDocumentVersionUpload', initiateDocumentVersionUpload_contentType - The content type of the document.

$sel:contentModifiedTimestamp:InitiateDocumentVersionUpload', initiateDocumentVersionUpload_contentModifiedTimestamp - The timestamp when the content of the document was modified.

$sel:id:InitiateDocumentVersionUpload', initiateDocumentVersionUpload_id - The ID of the document.

$sel:name:InitiateDocumentVersionUpload', initiateDocumentVersionUpload_name - The name of the document.

$sel:contentCreatedTimestamp:InitiateDocumentVersionUpload', initiateDocumentVersionUpload_contentCreatedTimestamp - The timestamp when the content of the document was originally created.

$sel:documentSizeInBytes:InitiateDocumentVersionUpload', initiateDocumentVersionUpload_documentSizeInBytes - The size of the document, in bytes.

$sel:authenticationToken:InitiateDocumentVersionUpload', initiateDocumentVersionUpload_authenticationToken - Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

$sel:parentFolderId:InitiateDocumentVersionUpload', initiateDocumentVersionUpload_parentFolderId - The ID of the parent folder.

Request Lenses

initiateDocumentVersionUpload_contentType :: Lens' InitiateDocumentVersionUpload (Maybe Text)

The content type of the document.

initiateDocumentVersionUpload_contentModifiedTimestamp :: Lens' InitiateDocumentVersionUpload (Maybe UTCTime)

The timestamp when the content of the document was modified.

initiateDocumentVersionUpload_id :: Lens' InitiateDocumentVersionUpload (Maybe Text)

The ID of the document.

initiateDocumentVersionUpload_name :: Lens' InitiateDocumentVersionUpload (Maybe Text)

The name of the document.

initiateDocumentVersionUpload_contentCreatedTimestamp :: Lens' InitiateDocumentVersionUpload (Maybe UTCTime)

The timestamp when the content of the document was originally created.

initiateDocumentVersionUpload_documentSizeInBytes :: Lens' InitiateDocumentVersionUpload (Maybe Integer)

The size of the document, in bytes.

initiateDocumentVersionUpload_authenticationToken :: Lens' InitiateDocumentVersionUpload (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

initiateDocumentVersionUpload_parentFolderId :: Lens' InitiateDocumentVersionUpload Text

The ID of the parent folder.

Destructuring the Response

data InitiateDocumentVersionUploadResponse

See: newInitiateDocumentVersionUploadResponse smart constructor.

Constructors

InitiateDocumentVersionUploadResponse'

Fields

uploadMetadata :: Maybe UploadMetadata

The upload metadata.

metadata :: Maybe DocumentMetadata

The document metadata.

httpStatus :: Int

The response's http status code.

newInitiateDocumentVersionUploadResponse

Arguments

:: Int
$sel:httpStatus:InitiateDocumentVersionUploadResponse'

-> InitiateDocumentVersionUploadResponse

Create a value of InitiateDocumentVersionUploadResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:uploadMetadata:InitiateDocumentVersionUploadResponse', initiateDocumentVersionUploadResponse_uploadMetadata - The upload metadata.

$sel:metadata:InitiateDocumentVersionUploadResponse', initiateDocumentVersionUploadResponse_metadata - The document metadata.

$sel:httpStatus:InitiateDocumentVersionUploadResponse', initiateDocumentVersionUploadResponse_httpStatus - The response's http status code.

Response Lenses

initiateDocumentVersionUploadResponse_uploadMetadata :: Lens' InitiateDocumentVersionUploadResponse (Maybe UploadMetadata)

The upload metadata.

initiateDocumentVersionUploadResponse_metadata :: Lens' InitiateDocumentVersionUploadResponse (Maybe DocumentMetadata)

The document metadata.

initiateDocumentVersionUploadResponse_httpStatus :: Lens' InitiateDocumentVersionUploadResponse Int

The response's http status code.