Network.AWS.WorkDocs.Lens

Description

Operations

DeleteCustomMetadata

deleteCustomMetadata_versionId :: Lens' DeleteCustomMetadata (Maybe Text)

The ID of the version, if the custom metadata is being deleted from a document version.

deleteCustomMetadata_keys :: Lens' DeleteCustomMetadata (Maybe [Text])

List of properties to remove.

deleteCustomMetadata_authenticationToken :: Lens' DeleteCustomMetadata (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

deleteCustomMetadata_deleteAll :: Lens' DeleteCustomMetadata (Maybe Bool)

Flag to indicate removal of all custom metadata properties from the specified resource.

deleteCustomMetadata_resourceId :: Lens' DeleteCustomMetadata Text

The ID of the resource, either a document or folder.

deleteCustomMetadataResponse_httpStatus :: Lens' DeleteCustomMetadataResponse Int

The response's http status code.

UpdateFolder

updateFolder_parentFolderId :: Lens' UpdateFolder (Maybe Text)

The ID of the parent folder.

updateFolder_name :: Lens' UpdateFolder (Maybe Text)

The name of the folder.

updateFolder_authenticationToken :: Lens' UpdateFolder (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

updateFolder_resourceState :: Lens' UpdateFolder (Maybe ResourceStateType)

The resource state of the folder. Only ACTIVE and RECYCLED are accepted values from the API.

updateFolder_folderId :: Lens' UpdateFolder Text

The ID of the folder.

DeleteNotificationSubscription

deleteNotificationSubscription_subscriptionId :: Lens' DeleteNotificationSubscription Text

The ID of the subscription.

deleteNotificationSubscription_organizationId :: Lens' DeleteNotificationSubscription Text

The ID of the organization.

DeleteFolder

deleteFolder_authenticationToken :: Lens' DeleteFolder (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

deleteFolder_folderId :: Lens' DeleteFolder Text

The ID of the folder.

DeleteLabels

deleteLabels_labels :: Lens' DeleteLabels (Maybe [Text])

List of labels to delete from the resource.

deleteLabels_authenticationToken :: Lens' DeleteLabels (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

deleteLabels_deleteAll :: Lens' DeleteLabels (Maybe Bool)

Flag to request removal of all labels from the specified resource.

deleteLabels_resourceId :: Lens' DeleteLabels Text

The ID of the resource.

deleteLabelsResponse_httpStatus :: Lens' DeleteLabelsResponse Int

The response's http status code.

AbortDocumentVersionUpload

abortDocumentVersionUpload_authenticationToken :: Lens' AbortDocumentVersionUpload (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

abortDocumentVersionUpload_documentId :: Lens' AbortDocumentVersionUpload Text

The ID of the document.

abortDocumentVersionUpload_versionId :: Lens' AbortDocumentVersionUpload Text

The ID of the version.

UpdateDocumentVersion

updateDocumentVersion_versionStatus :: Lens' UpdateDocumentVersion (Maybe DocumentVersionStatus)

The status of the version.

updateDocumentVersion_authenticationToken :: Lens' UpdateDocumentVersion (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

updateDocumentVersion_documentId :: Lens' UpdateDocumentVersion Text

The ID of the document.

updateDocumentVersion_versionId :: Lens' UpdateDocumentVersion Text

The version ID of the document.

DeactivateUser

deactivateUser_authenticationToken :: Lens' DeactivateUser (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

deactivateUser_userId :: Lens' DeactivateUser Text

The ID of the user.

DescribeFolderContents

describeFolderContents_include :: Lens' DescribeFolderContents (Maybe Text)

The contents to include. Specify "INITIALIZED" to include initialized documents.

describeFolderContents_order :: Lens' DescribeFolderContents (Maybe OrderType)

The order for the contents of the folder.

describeFolderContents_authenticationToken :: Lens' DescribeFolderContents (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

describeFolderContents_type :: Lens' DescribeFolderContents (Maybe FolderContentType)

The type of items.

describeFolderContents_limit :: Lens' DescribeFolderContents (Maybe Natural)

The maximum number of items to return with this call.

describeFolderContents_sort :: Lens' DescribeFolderContents (Maybe ResourceSortType)

The sorting criteria.

describeFolderContents_marker :: Lens' DescribeFolderContents (Maybe Text)

The marker for the next set of results. This marker was received from a previous call.

describeFolderContents_folderId :: Lens' DescribeFolderContents Text

The ID of the folder.

describeFolderContentsResponse_documents :: Lens' DescribeFolderContentsResponse (Maybe [DocumentMetadata])

The documents in the specified folder.

describeFolderContentsResponse_folders :: Lens' DescribeFolderContentsResponse (Maybe [FolderMetadata])

The subfolders in the specified folder.

describeFolderContentsResponse_marker :: Lens' DescribeFolderContentsResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

describeFolderContentsResponse_httpStatus :: Lens' DescribeFolderContentsResponse Int

The response's http status code.

CreateLabels

createLabels_authenticationToken :: Lens' CreateLabels (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

createLabels_resourceId :: Lens' CreateLabels Text

The ID of the resource.

createLabels_labels :: Lens' CreateLabels [Text]

List of labels to add to the resource.

createLabelsResponse_httpStatus :: Lens' CreateLabelsResponse Int

The response's http status code.

DescribeRootFolders

describeRootFolders_limit :: Lens' DescribeRootFolders (Maybe Natural)

The maximum number of items to return.

describeRootFolders_marker :: Lens' DescribeRootFolders (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

describeRootFolders_authenticationToken :: Lens' DescribeRootFolders Text

Amazon WorkDocs authentication token.

describeRootFoldersResponse_folders :: Lens' DescribeRootFoldersResponse (Maybe [FolderMetadata])

The user's special folders.

describeRootFoldersResponse_marker :: Lens' DescribeRootFoldersResponse (Maybe Text)

The marker for the next set of results.

describeRootFoldersResponse_httpStatus :: Lens' DescribeRootFoldersResponse Int

The response's http status code.

UpdateDocument

updateDocument_parentFolderId :: Lens' UpdateDocument (Maybe Text)

The ID of the parent folder.

updateDocument_name :: Lens' UpdateDocument (Maybe Text)

The name of the document.

updateDocument_authenticationToken :: Lens' UpdateDocument (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

updateDocument_resourceState :: Lens' UpdateDocument (Maybe ResourceStateType)

The resource state of the document. Only ACTIVE and RECYCLED are supported.

updateDocument_documentId :: Lens' UpdateDocument Text

The ID of the document.

DeleteDocument

deleteDocument_authenticationToken :: Lens' DeleteDocument (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

deleteDocument_documentId :: Lens' DeleteDocument Text

The ID of the document.

DescribeDocumentVersions

describeDocumentVersions_include :: Lens' DescribeDocumentVersions (Maybe Text)

A comma-separated list of values. Specify "INITIALIZED" to include incomplete versions.

describeDocumentVersions_fields :: Lens' DescribeDocumentVersions (Maybe Text)

Specify "SOURCE" to include initialized versions and a URL for the source document.

describeDocumentVersions_authenticationToken :: Lens' DescribeDocumentVersions (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

describeDocumentVersions_limit :: Lens' DescribeDocumentVersions (Maybe Natural)

The maximum number of versions to return with this call.

describeDocumentVersions_marker :: Lens' DescribeDocumentVersions (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

describeDocumentVersions_documentId :: Lens' DescribeDocumentVersions Text

The ID of the document.

describeDocumentVersionsResponse_documentVersions :: Lens' DescribeDocumentVersionsResponse (Maybe [DocumentVersionMetadata])

The document versions.

describeDocumentVersionsResponse_marker :: Lens' DescribeDocumentVersionsResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

describeDocumentVersionsResponse_httpStatus :: Lens' DescribeDocumentVersionsResponse Int

The response's http status code.

GetDocumentVersion

getDocumentVersion_includeCustomMetadata :: Lens' GetDocumentVersion (Maybe Bool)

Set this to TRUE to include custom metadata in the response.

getDocumentVersion_fields :: Lens' GetDocumentVersion (Maybe Text)

A comma-separated list of values. Specify "SOURCE" to include a URL for the source document.

getDocumentVersion_authenticationToken :: Lens' GetDocumentVersion (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

getDocumentVersion_documentId :: Lens' GetDocumentVersion Text

The ID of the document.

getDocumentVersion_versionId :: Lens' GetDocumentVersion Text

The version ID of the document.

getDocumentVersionResponse_metadata :: Lens' GetDocumentVersionResponse (Maybe DocumentVersionMetadata)

The version metadata.

getDocumentVersionResponse_customMetadata :: Lens' GetDocumentVersionResponse (Maybe (HashMap Text Text))

The custom metadata on the document version.

getDocumentVersionResponse_httpStatus :: Lens' GetDocumentVersionResponse Int

The response's http status code.

ActivateUser

activateUser_authenticationToken :: Lens' ActivateUser (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

activateUser_userId :: Lens' ActivateUser Text

The ID of the user.

activateUserResponse_user :: Lens' ActivateUserResponse (Maybe User)

The user information.

activateUserResponse_httpStatus :: Lens' ActivateUserResponse Int

The response's http status code.

GetFolderPath

getFolderPath_fields :: Lens' GetFolderPath (Maybe Text)

A comma-separated list of values. Specify "NAME" to include the names of the parent folders.

getFolderPath_authenticationToken :: Lens' GetFolderPath (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

getFolderPath_limit :: Lens' GetFolderPath (Maybe Natural)

The maximum number of levels in the hierarchy to return.

getFolderPath_marker :: Lens' GetFolderPath (Maybe Text)

This value is not supported.

getFolderPath_folderId :: Lens' GetFolderPath Text

The ID of the folder.

getFolderPathResponse_path :: Lens' GetFolderPathResponse (Maybe ResourcePath)

The path information.

getFolderPathResponse_httpStatus :: Lens' GetFolderPathResponse Int

The response's http status code.

CreateUser

createUser_storageRule :: Lens' CreateUser (Maybe StorageRuleType)

The amount of storage for the user.

createUser_organizationId :: Lens' CreateUser (Maybe Text)

The ID of the organization.

createUser_timeZoneId :: Lens' CreateUser (Maybe Text)

The time zone ID of the user.

createUser_authenticationToken :: Lens' CreateUser (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

createUser_emailAddress :: Lens' CreateUser (Maybe Text)

The email address of the user.

createUser_username :: Lens' CreateUser Text

The login name of the user.

createUser_givenName :: Lens' CreateUser Text

The given name of the user.

createUser_surname :: Lens' CreateUser Text

The surname of the user.

createUser_password :: Lens' CreateUser Text

The password of the user.

createUserResponse_user :: Lens' CreateUserResponse (Maybe User)

The user information.

createUserResponse_httpStatus :: Lens' CreateUserResponse Int

The response's http status code.

DeleteComment

deleteComment_authenticationToken :: Lens' DeleteComment (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

deleteComment_documentId :: Lens' DeleteComment Text

The ID of the document.

deleteComment_versionId :: Lens' DeleteComment Text

The ID of the document version.

deleteComment_commentId :: Lens' DeleteComment Text

The ID of the comment.

CreateFolder

createFolder_name :: Lens' CreateFolder (Maybe Text)

The name of the new folder.

createFolder_authenticationToken :: Lens' CreateFolder (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

createFolder_parentFolderId :: Lens' CreateFolder Text

The ID of the parent folder.

createFolderResponse_metadata :: Lens' CreateFolderResponse (Maybe FolderMetadata)

The metadata of the folder.

createFolderResponse_httpStatus :: Lens' CreateFolderResponse Int

The response's http status code.

CreateCustomMetadata

createCustomMetadata_versionId :: Lens' CreateCustomMetadata (Maybe Text)

The ID of the version, if the custom metadata is being added to a document version.

createCustomMetadata_authenticationToken :: Lens' CreateCustomMetadata (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

createCustomMetadata_resourceId :: Lens' CreateCustomMetadata Text

The ID of the resource.

createCustomMetadata_customMetadata :: Lens' CreateCustomMetadata (HashMap Text Text)

Custom metadata in the form of name-value pairs.

createCustomMetadataResponse_httpStatus :: Lens' CreateCustomMetadataResponse Int

The response's http status code.

CreateNotificationSubscription

createNotificationSubscription_organizationId :: Lens' CreateNotificationSubscription Text

The ID of the organization.

createNotificationSubscription_endpoint :: Lens' CreateNotificationSubscription Text

The endpoint to receive the notifications. If the protocol is HTTPS, the endpoint is a URL that begins with https.

createNotificationSubscription_protocol :: Lens' CreateNotificationSubscription SubscriptionProtocolType

The protocol to use. The supported value is https, which delivers JSON-encoded messages using HTTPS POST.

createNotificationSubscription_subscriptionType :: Lens' CreateNotificationSubscription SubscriptionType

The notification type.

createNotificationSubscriptionResponse_subscription :: Lens' CreateNotificationSubscriptionResponse (Maybe Subscription)

The subscription.

createNotificationSubscriptionResponse_httpStatus :: Lens' CreateNotificationSubscriptionResponse Int

The response's http status code.

GetResources

getResources_collectionType :: Lens' GetResources (Maybe ResourceCollectionType)

The collection type.

getResources_userId :: Lens' GetResources (Maybe Text)

The user ID for the resource collection. This is a required field for accessing the API operation using IAM credentials.

getResources_authenticationToken :: Lens' GetResources (Maybe Text)

The Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

getResources_limit :: Lens' GetResources (Maybe Natural)

The maximum number of resources to return.

getResources_marker :: Lens' GetResources (Maybe Text)

The marker for the next set of results. This marker was received from a previous call.

getResourcesResponse_documents :: Lens' GetResourcesResponse (Maybe [DocumentMetadata])

The documents in the specified collection.

getResourcesResponse_folders :: Lens' GetResourcesResponse (Maybe [FolderMetadata])

The folders in the specified folder.

getResourcesResponse_marker :: Lens' GetResourcesResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

getResourcesResponse_httpStatus :: Lens' GetResourcesResponse Int

The response's http status code.

RemoveResourcePermission

removeResourcePermission_authenticationToken :: Lens' RemoveResourcePermission (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

removeResourcePermission_principalType :: Lens' RemoveResourcePermission (Maybe PrincipalType)

The principal type of the resource.

removeResourcePermission_resourceId :: Lens' RemoveResourcePermission Text

The ID of the resource.

removeResourcePermission_principalId :: Lens' RemoveResourcePermission Text

The principal ID of the resource.

CreateComment

createComment_parentId :: Lens' CreateComment (Maybe Text)

The ID of the parent comment.

createComment_visibility :: Lens' CreateComment (Maybe CommentVisibilityType)

The visibility of the comment. Options are either PRIVATE, where the comment is visible only to the comment author and document owner and co-owners, or PUBLIC, where the comment is visible to document owners, co-owners, and contributors.

createComment_threadId :: Lens' CreateComment (Maybe Text)

The ID of the root comment in the thread.

createComment_authenticationToken :: Lens' CreateComment (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

createComment_notifyCollaborators :: Lens' CreateComment (Maybe Bool)

Set this parameter to TRUE to send an email out to the document collaborators after the comment is created.

createComment_documentId :: Lens' CreateComment Text

The ID of the document.

createComment_versionId :: Lens' CreateComment Text

The ID of the document version.

createComment_text :: Lens' CreateComment Text

The text of the comment.

createCommentResponse_comment :: Lens' CreateCommentResponse (Maybe Comment)

The comment that has been created.

createCommentResponse_httpStatus :: Lens' CreateCommentResponse Int

The response's http status code.

DescribeResourcePermissions

describeResourcePermissions_principalId :: Lens' DescribeResourcePermissions (Maybe Text)

The ID of the principal to filter permissions by.

describeResourcePermissions_authenticationToken :: Lens' DescribeResourcePermissions (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

describeResourcePermissions_limit :: Lens' DescribeResourcePermissions (Maybe Natural)

The maximum number of items to return with this call.

describeResourcePermissions_marker :: Lens' DescribeResourcePermissions (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call)

describeResourcePermissions_resourceId :: Lens' DescribeResourcePermissions Text

The ID of the resource.

describeResourcePermissionsResponse_principals :: Lens' DescribeResourcePermissionsResponse (Maybe [Principal])

The principals.

describeResourcePermissionsResponse_marker :: Lens' DescribeResourcePermissionsResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

describeResourcePermissionsResponse_httpStatus :: Lens' DescribeResourcePermissionsResponse Int

The response's http status code.

DescribeUsers

describeUsers_organizationId :: Lens' DescribeUsers (Maybe Text)

The ID of the organization.

describeUsers_query :: Lens' DescribeUsers (Maybe Text)

A query to filter users by user name.

describeUsers_userIds :: Lens' DescribeUsers (Maybe Text)

The IDs of the users.

describeUsers_include :: Lens' DescribeUsers (Maybe UserFilterType)

The state of the users. Specify "ALL" to include inactive users.

describeUsers_fields :: Lens' DescribeUsers (Maybe Text)

A comma-separated list of values. Specify "STORAGE_METADATA" to include the user storage quota and utilization information.

describeUsers_order :: Lens' DescribeUsers (Maybe OrderType)

The order for the results.

describeUsers_authenticationToken :: Lens' DescribeUsers (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

describeUsers_limit :: Lens' DescribeUsers (Maybe Natural)

The maximum number of items to return.

describeUsers_sort :: Lens' DescribeUsers (Maybe UserSortType)

The sorting criteria.

describeUsers_marker :: Lens' DescribeUsers (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

describeUsersResponse_totalNumberOfUsers :: Lens' DescribeUsersResponse (Maybe Integer)

The total number of users included in the results.

describeUsersResponse_users :: Lens' DescribeUsersResponse (Maybe [User])

The users.

describeUsersResponse_marker :: Lens' DescribeUsersResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

describeUsersResponse_httpStatus :: Lens' DescribeUsersResponse Int

The response's http status code.

GetDocumentPath

getDocumentPath_fields :: Lens' GetDocumentPath (Maybe Text)

A comma-separated list of values. Specify NAME to include the names of the parent folders.

getDocumentPath_authenticationToken :: Lens' GetDocumentPath (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

getDocumentPath_limit :: Lens' GetDocumentPath (Maybe Natural)

The maximum number of levels in the hierarchy to return.

getDocumentPath_marker :: Lens' GetDocumentPath (Maybe Text)

This value is not supported.

getDocumentPath_documentId :: Lens' GetDocumentPath Text

The ID of the document.

getDocumentPathResponse_path :: Lens' GetDocumentPathResponse (Maybe ResourcePath)

The path information.

getDocumentPathResponse_httpStatus :: Lens' GetDocumentPathResponse Int

The response's http status code.

GetDocument

getDocument_includeCustomMetadata :: Lens' GetDocument (Maybe Bool)

Set this to TRUE to include custom metadata in the response.

getDocument_authenticationToken :: Lens' GetDocument (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

getDocument_documentId :: Lens' GetDocument Text

The ID of the document.

getDocumentResponse_metadata :: Lens' GetDocumentResponse (Maybe DocumentMetadata)

The metadata details of the document.

getDocumentResponse_customMetadata :: Lens' GetDocumentResponse (Maybe (HashMap Text Text))

The custom metadata on the document.

getDocumentResponse_httpStatus :: Lens' GetDocumentResponse Int

The response's http status code.

DescribeGroups

describeGroups_organizationId :: Lens' DescribeGroups (Maybe Text)

The ID of the organization.

describeGroups_authenticationToken :: Lens' DescribeGroups (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

describeGroups_limit :: Lens' DescribeGroups (Maybe Natural)

The maximum number of items to return with this call.

describeGroups_marker :: Lens' DescribeGroups (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

describeGroups_searchQuery :: Lens' DescribeGroups Text

A query to describe groups by group name.

describeGroupsResponse_groups :: Lens' DescribeGroupsResponse (Maybe [GroupMetadata])

The list of groups.

describeGroupsResponse_marker :: Lens' DescribeGroupsResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

describeGroupsResponse_httpStatus :: Lens' DescribeGroupsResponse Int

The response's http status code.

DescribeActivities

describeActivities_resourceId :: Lens' DescribeActivities (Maybe Text)

The document or folder ID for which to describe activity types.

describeActivities_organizationId :: Lens' DescribeActivities (Maybe Text)

The ID of the organization. This is a mandatory parameter when using administrative API (SigV4) requests.

describeActivities_includeIndirectActivities :: Lens' DescribeActivities (Maybe Bool)

Includes indirect activities. An indirect activity results from a direct activity performed on a parent resource. For example, sharing a parent folder (the direct activity) shares all of the subfolders and documents within the parent folder (the indirect activity).

describeActivities_startTime :: Lens' DescribeActivities (Maybe UTCTime)

The timestamp that determines the starting time of the activities. The response includes the activities performed after the specified timestamp.

describeActivities_endTime :: Lens' DescribeActivities (Maybe UTCTime)

The timestamp that determines the end time of the activities. The response includes the activities performed before the specified timestamp.

describeActivities_userId :: Lens' DescribeActivities (Maybe Text)

The ID of the user who performed the action. The response includes activities pertaining to this user. This is an optional parameter and is only applicable for administrative API (SigV4) requests.

describeActivities_activityTypes :: Lens' DescribeActivities (Maybe Text)

Specifies which activity types to include in the response. If this field is left empty, all activity types are returned.

describeActivities_authenticationToken :: Lens' DescribeActivities (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

describeActivities_limit :: Lens' DescribeActivities (Maybe Natural)

The maximum number of items to return.

describeActivities_marker :: Lens' DescribeActivities (Maybe Text)

The marker for the next set of results.

describeActivitiesResponse_userActivities :: Lens' DescribeActivitiesResponse (Maybe [Activity])

The list of activities for the specified user and time period.

describeActivitiesResponse_marker :: Lens' DescribeActivitiesResponse (Maybe Text)

The marker for the next set of results.

describeActivitiesResponse_httpStatus :: Lens' DescribeActivitiesResponse Int

The response's http status code.

GetCurrentUser

getCurrentUser_authenticationToken :: Lens' GetCurrentUser Text

Amazon WorkDocs authentication token.

getCurrentUserResponse_user :: Lens' GetCurrentUserResponse (Maybe User)

Metadata of the user.

getCurrentUserResponse_httpStatus :: Lens' GetCurrentUserResponse Int

The response's http status code.

AddResourcePermissions

addResourcePermissions_notificationOptions :: Lens' AddResourcePermissions (Maybe NotificationOptions)

The notification options.

addResourcePermissions_authenticationToken :: Lens' AddResourcePermissions (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

addResourcePermissions_resourceId :: Lens' AddResourcePermissions Text

The ID of the resource.

addResourcePermissions_principals :: Lens' AddResourcePermissions [SharePrincipal]

The users, groups, or organization being granted permission.

addResourcePermissionsResponse_shareResults :: Lens' AddResourcePermissionsResponse (Maybe [ShareResult])

The share results.

addResourcePermissionsResponse_httpStatus :: Lens' AddResourcePermissionsResponse Int

The response's http status code.

UpdateUser

updateUser_storageRule :: Lens' UpdateUser (Maybe StorageRuleType)

The amount of storage for the user.

updateUser_grantPoweruserPrivileges :: Lens' UpdateUser (Maybe BooleanEnumType)

Boolean value to determine whether the user is granted Poweruser privileges.

updateUser_timeZoneId :: Lens' UpdateUser (Maybe Text)

The time zone ID of the user.

updateUser_surname :: Lens' UpdateUser (Maybe Text)

The surname of the user.

updateUser_locale :: Lens' UpdateUser (Maybe LocaleType)

The locale of the user.

updateUser_givenName :: Lens' UpdateUser (Maybe Text)

The given name of the user.

updateUser_authenticationToken :: Lens' UpdateUser (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

updateUser_type :: Lens' UpdateUser (Maybe UserType)

The type of the user.

updateUser_userId :: Lens' UpdateUser Text

The ID of the user.

updateUserResponse_user :: Lens' UpdateUserResponse (Maybe User)

The user information.

updateUserResponse_httpStatus :: Lens' UpdateUserResponse Int

The response's http status code.

DescribeNotificationSubscriptions

describeNotificationSubscriptions_limit :: Lens' DescribeNotificationSubscriptions (Maybe Natural)

The maximum number of items to return with this call.

describeNotificationSubscriptions_marker :: Lens' DescribeNotificationSubscriptions (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

describeNotificationSubscriptions_organizationId :: Lens' DescribeNotificationSubscriptions Text

The ID of the organization.

describeNotificationSubscriptionsResponse_subscriptions :: Lens' DescribeNotificationSubscriptionsResponse (Maybe [Subscription])

The subscriptions.

describeNotificationSubscriptionsResponse_marker :: Lens' DescribeNotificationSubscriptionsResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

describeNotificationSubscriptionsResponse_httpStatus :: Lens' DescribeNotificationSubscriptionsResponse Int

The response's http status code.

GetFolder

getFolder_includeCustomMetadata :: Lens' GetFolder (Maybe Bool)

Set to TRUE to include custom metadata in the response.

getFolder_authenticationToken :: Lens' GetFolder (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

getFolder_folderId :: Lens' GetFolder Text

The ID of the folder.

getFolderResponse_metadata :: Lens' GetFolderResponse (Maybe FolderMetadata)

The metadata of the folder.

getFolderResponse_customMetadata :: Lens' GetFolderResponse (Maybe (HashMap Text Text))

The custom metadata on the folder.

getFolderResponse_httpStatus :: Lens' GetFolderResponse Int

The response's http status code.

DeleteUser

deleteUser_authenticationToken :: Lens' DeleteUser (Maybe Text)

Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.

deleteUser_userId :: Lens' DeleteUser Text

The ID of the user.

DeleteFolderContents

deleteFolderContents_authenticationToken :: Lens' DeleteFolderContents (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

deleteFolderContents_folderId :: Lens' DeleteFolderContents Text

The ID of the folder.

DescribeComments

describeComments_authenticationToken :: Lens' DescribeComments (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

describeComments_limit :: Lens' DescribeComments (Maybe Natural)

The maximum number of items to return.

describeComments_marker :: Lens' DescribeComments (Maybe Text)

The marker for the next set of results. This marker was received from a previous call.

describeComments_documentId :: Lens' DescribeComments Text

The ID of the document.

describeComments_versionId :: Lens' DescribeComments Text

The ID of the document version.

describeCommentsResponse_comments :: Lens' DescribeCommentsResponse (Maybe [Comment])

The list of comments for the specified document version.

describeCommentsResponse_marker :: Lens' DescribeCommentsResponse (Maybe Text)

The marker for the next set of results. This marker was received from a previous call.

describeCommentsResponse_httpStatus :: Lens' DescribeCommentsResponse Int

The response's http status code.

InitiateDocumentVersionUpload

initiateDocumentVersionUpload_contentType :: Lens' InitiateDocumentVersionUpload (Maybe Text)

The content type of the document.

initiateDocumentVersionUpload_contentModifiedTimestamp :: Lens' InitiateDocumentVersionUpload (Maybe UTCTime)

The timestamp when the content of the document was modified.

initiateDocumentVersionUpload_id :: Lens' InitiateDocumentVersionUpload (Maybe Text)

The ID of the document.

initiateDocumentVersionUpload_name :: Lens' InitiateDocumentVersionUpload (Maybe Text)

The name of the document.

initiateDocumentVersionUpload_contentCreatedTimestamp :: Lens' InitiateDocumentVersionUpload (Maybe UTCTime)

The timestamp when the content of the document was originally created.

initiateDocumentVersionUpload_documentSizeInBytes :: Lens' InitiateDocumentVersionUpload (Maybe Integer)

The size of the document, in bytes.

initiateDocumentVersionUpload_authenticationToken :: Lens' InitiateDocumentVersionUpload (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

initiateDocumentVersionUpload_parentFolderId :: Lens' InitiateDocumentVersionUpload Text

The ID of the parent folder.

initiateDocumentVersionUploadResponse_uploadMetadata :: Lens' InitiateDocumentVersionUploadResponse (Maybe UploadMetadata)

The upload metadata.

initiateDocumentVersionUploadResponse_metadata :: Lens' InitiateDocumentVersionUploadResponse (Maybe DocumentMetadata)

The document metadata.

initiateDocumentVersionUploadResponse_httpStatus :: Lens' InitiateDocumentVersionUploadResponse Int

The response's http status code.

RemoveAllResourcePermissions

removeAllResourcePermissions_authenticationToken :: Lens' RemoveAllResourcePermissions (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

removeAllResourcePermissions_resourceId :: Lens' RemoveAllResourcePermissions Text

The ID of the resource.

Types

Activity

activity_resourceMetadata :: Lens' Activity (Maybe ResourceMetadata)

The metadata of the resource involved in the user action.

activity_organizationId :: Lens' Activity (Maybe Text)

The ID of the organization.

activity_originalParent :: Lens' Activity (Maybe ResourceMetadata)

The original parent of the resource. This is an optional field and is filled for move activities.

activity_participants :: Lens' Activity (Maybe Participants)

The list of users or groups impacted by this action. This is an optional field and is filled for the following sharing activities: DOCUMENT_SHARED, DOCUMENT_SHARED, DOCUMENT_UNSHARED, FOLDER_SHARED, FOLDER_UNSHARED.

activity_commentMetadata :: Lens' Activity (Maybe CommentMetadata)

Metadata of the commenting activity. This is an optional field and is filled for commenting activities.

activity_timeStamp :: Lens' Activity (Maybe UTCTime)

The timestamp when the action was performed.

activity_initiator :: Lens' Activity (Maybe UserMetadata)

The user who performed the action.

activity_type :: Lens' Activity (Maybe ActivityType)

The activity type.

activity_isIndirectActivity :: Lens' Activity (Maybe Bool)

Indicates whether an activity is indirect or direct. An indirect activity results from a direct activity performed on a parent resource. For example, sharing a parent folder (the direct activity) shares all of the subfolders and documents within the parent folder (the indirect activity).

Comment

comment_status :: Lens' Comment (Maybe CommentStatusType)

The status of the comment.

comment_createdTimestamp :: Lens' Comment (Maybe UTCTime)

The time that the comment was created.

comment_contributor :: Lens' Comment (Maybe User)

The details of the user who made the comment.

comment_parentId :: Lens' Comment (Maybe Text)

The ID of the parent comment.

comment_recipientId :: Lens' Comment (Maybe Text)

If the comment is a reply to another user's comment, this field contains the user ID of the user being replied to.

comment_visibility :: Lens' Comment (Maybe CommentVisibilityType)

The visibility of the comment. Options are either PRIVATE, where the comment is visible only to the comment author and document owner and co-owners, or PUBLIC, where the comment is visible to document owners, co-owners, and contributors.

comment_threadId :: Lens' Comment (Maybe Text)

The ID of the root comment in the thread.

comment_text :: Lens' Comment (Maybe Text)

The text of the comment.

comment_commentId :: Lens' Comment Text

The ID of the comment.

CommentMetadata

commentMetadata_commentStatus :: Lens' CommentMetadata (Maybe CommentStatusType)

The status of the comment.

commentMetadata_createdTimestamp :: Lens' CommentMetadata (Maybe UTCTime)

The timestamp that the comment was created.

commentMetadata_contributor :: Lens' CommentMetadata (Maybe User)

The user who made the comment.

commentMetadata_recipientId :: Lens' CommentMetadata (Maybe Text)

The ID of the user being replied to.

commentMetadata_commentId :: Lens' CommentMetadata (Maybe Text)

The ID of the comment.

DocumentMetadata

documentMetadata_modifiedTimestamp :: Lens' DocumentMetadata (Maybe UTCTime)

The time when the document was updated.

documentMetadata_parentFolderId :: Lens' DocumentMetadata (Maybe Text)

The ID of the parent folder.

documentMetadata_creatorId :: Lens' DocumentMetadata (Maybe Text)

The ID of the creator.

documentMetadata_createdTimestamp :: Lens' DocumentMetadata (Maybe UTCTime)

The time when the document was created.

documentMetadata_labels :: Lens' DocumentMetadata (Maybe [Text])

List of labels on the document.

documentMetadata_id :: Lens' DocumentMetadata (Maybe Text)

The ID of the document.

documentMetadata_latestVersionMetadata :: Lens' DocumentMetadata (Maybe DocumentVersionMetadata)

The latest version of the document.

documentMetadata_resourceState :: Lens' DocumentMetadata (Maybe ResourceStateType)

The resource state.

DocumentVersionMetadata

documentVersionMetadata_modifiedTimestamp :: Lens' DocumentVersionMetadata (Maybe UTCTime)

The timestamp when the document was last uploaded.

documentVersionMetadata_status :: Lens' DocumentVersionMetadata (Maybe DocumentStatusType)

The status of the document.

documentVersionMetadata_creatorId :: Lens' DocumentVersionMetadata (Maybe Text)

The ID of the creator.

documentVersionMetadata_contentType :: Lens' DocumentVersionMetadata (Maybe Text)

The content type of the document.

documentVersionMetadata_contentModifiedTimestamp :: Lens' DocumentVersionMetadata (Maybe UTCTime)

The timestamp when the content of the document was modified.

documentVersionMetadata_createdTimestamp :: Lens' DocumentVersionMetadata (Maybe UTCTime)

The timestamp when the document was first uploaded.

documentVersionMetadata_id :: Lens' DocumentVersionMetadata (Maybe Text)

The ID of the version.

documentVersionMetadata_source :: Lens' DocumentVersionMetadata (Maybe (HashMap DocumentSourceType Text))

The source of the document.

documentVersionMetadata_name :: Lens' DocumentVersionMetadata (Maybe Text)

The name of the version.

documentVersionMetadata_contentCreatedTimestamp :: Lens' DocumentVersionMetadata (Maybe UTCTime)

The timestamp when the content of the document was originally created.

documentVersionMetadata_signature :: Lens' DocumentVersionMetadata (Maybe Text)

The signature of the document.

documentVersionMetadata_thumbnail :: Lens' DocumentVersionMetadata (Maybe (HashMap DocumentThumbnailType Text))

The thumbnail of the document.

documentVersionMetadata_size :: Lens' DocumentVersionMetadata (Maybe Integer)

The size of the document, in bytes.

FolderMetadata

folderMetadata_modifiedTimestamp :: Lens' FolderMetadata (Maybe UTCTime)

The time when the folder was updated.

folderMetadata_parentFolderId :: Lens' FolderMetadata (Maybe Text)

The ID of the parent folder.

folderMetadata_latestVersionSize :: Lens' FolderMetadata (Maybe Integer)

The size of the latest version of the folder metadata.

folderMetadata_creatorId :: Lens' FolderMetadata (Maybe Text)

The ID of the creator.

folderMetadata_createdTimestamp :: Lens' FolderMetadata (Maybe UTCTime)

The time when the folder was created.

folderMetadata_labels :: Lens' FolderMetadata (Maybe [Text])

List of labels on the folder.

folderMetadata_id :: Lens' FolderMetadata (Maybe Text)

The ID of the folder.

folderMetadata_name :: Lens' FolderMetadata (Maybe Text)

The name of the folder.

folderMetadata_signature :: Lens' FolderMetadata (Maybe Text)

The unique identifier created from the subfolders and documents of the folder.

folderMetadata_size :: Lens' FolderMetadata (Maybe Integer)

The size of the folder metadata.

folderMetadata_resourceState :: Lens' FolderMetadata (Maybe ResourceStateType)

The resource state of the folder.

GroupMetadata

groupMetadata_id :: Lens' GroupMetadata (Maybe Text)

The ID of the user group.

groupMetadata_name :: Lens' GroupMetadata (Maybe Text)

The name of the group.

NotificationOptions

notificationOptions_sendEmail :: Lens' NotificationOptions (Maybe Bool)

Boolean value to indicate an email notification should be sent to the receipients.

notificationOptions_emailMessage :: Lens' NotificationOptions (Maybe Text)

Text value to be included in the email body.

Participants

participants_groups :: Lens' Participants (Maybe [GroupMetadata])

The list of user groups.

participants_users :: Lens' Participants (Maybe [UserMetadata])

The list of users.

PermissionInfo

permissionInfo_role :: Lens' PermissionInfo (Maybe RoleType)

The role of the user.

permissionInfo_type :: Lens' PermissionInfo (Maybe RolePermissionType)

The type of permissions.

Principal

principal_id :: Lens' Principal (Maybe Text)

The ID of the resource.

principal_roles :: Lens' Principal (Maybe [PermissionInfo])

The permission information for the resource.

principal_type :: Lens' Principal (Maybe PrincipalType)

The type of resource.

ResourceMetadata

resourceMetadata_originalName :: Lens' ResourceMetadata (Maybe Text)

The original name of the resource before a rename operation.

resourceMetadata_id :: Lens' ResourceMetadata (Maybe Text)

The ID of the resource.

resourceMetadata_name :: Lens' ResourceMetadata (Maybe Text)

The name of the resource.

resourceMetadata_versionId :: Lens' ResourceMetadata (Maybe Text)

The version ID of the resource. This is an optional field and is filled for action on document version.

resourceMetadata_owner :: Lens' ResourceMetadata (Maybe UserMetadata)

The owner of the resource.

resourceMetadata_parentId :: Lens' ResourceMetadata (Maybe Text)

The parent ID of the resource before a rename operation.

resourceMetadata_type :: Lens' ResourceMetadata (Maybe ResourceType)

The type of resource.

ResourcePath

resourcePath_components :: Lens' ResourcePath (Maybe [ResourcePathComponent])

The components of the resource path.

ResourcePathComponent

resourcePathComponent_id :: Lens' ResourcePathComponent (Maybe Text)

The ID of the resource path.

resourcePathComponent_name :: Lens' ResourcePathComponent (Maybe Text)

The name of the resource path.

SharePrincipal

sharePrincipal_id :: Lens' SharePrincipal Text

The ID of the recipient.

sharePrincipal_type :: Lens' SharePrincipal PrincipalType

The type of the recipient.

sharePrincipal_role :: Lens' SharePrincipal RoleType

The role of the recipient.

ShareResult

shareResult_statusMessage :: Lens' ShareResult (Maybe Text)

The status message.

shareResult_status :: Lens' ShareResult (Maybe ShareStatusType)

The status.

shareResult_inviteePrincipalId :: Lens' ShareResult (Maybe Text)

The ID of the invited user.

shareResult_shareId :: Lens' ShareResult (Maybe Text)

The ID of the resource that was shared.

shareResult_principalId :: Lens' ShareResult (Maybe Text)

The ID of the principal.

shareResult_role :: Lens' ShareResult (Maybe RoleType)

The role.

StorageRuleType

storageRuleType_storageType :: Lens' StorageRuleType (Maybe StorageType)

The type of storage.

storageRuleType_storageAllocatedInBytes :: Lens' StorageRuleType (Maybe Natural)

The amount of storage allocated, in bytes.

Subscription

subscription_subscriptionId :: Lens' Subscription (Maybe Text)

The ID of the subscription.

subscription_protocol :: Lens' Subscription (Maybe SubscriptionProtocolType)

The protocol of the subscription.

subscription_endPoint :: Lens' Subscription (Maybe Text)

The endpoint of the subscription.

UploadMetadata

uploadMetadata_signedHeaders :: Lens' UploadMetadata (Maybe (HashMap Text Text))

The signed headers.

uploadMetadata_uploadUrl :: Lens' UploadMetadata (Maybe Text)

The URL of the upload.

User

user_modifiedTimestamp :: Lens' User (Maybe UTCTime)

The time when the user was modified.

user_status :: Lens' User (Maybe UserStatusType)

The status of the user.

user_organizationId :: Lens' User (Maybe Text)

The ID of the organization.

user_timeZoneId :: Lens' User (Maybe Text)

The time zone ID of the user.

user_createdTimestamp :: Lens' User (Maybe UTCTime)

The time when the user was created.

user_surname :: Lens' User (Maybe Text)

The surname of the user.

user_locale :: Lens' User (Maybe LocaleType)

The locale of the user.

user_rootFolderId :: Lens' User (Maybe Text)

The ID of the root folder.

user_id :: Lens' User (Maybe Text)

The ID of the user.

user_givenName :: Lens' User (Maybe Text)

The given name of the user.

user_recycleBinFolderId :: Lens' User (Maybe Text)

The ID of the recycle bin folder.

user_storage :: Lens' User (Maybe UserStorageMetadata)

The storage for the user.

user_username :: Lens' User (Maybe Text)

The login name of the user.

user_type :: Lens' User (Maybe UserType)

The type of user.

user_emailAddress :: Lens' User (Maybe Text)

The email address of the user.

UserMetadata

userMetadata_surname :: Lens' UserMetadata (Maybe Text)

The surname of the user.

userMetadata_id :: Lens' UserMetadata (Maybe Text)

The ID of the user.

userMetadata_givenName :: Lens' UserMetadata (Maybe Text)

The given name of the user before a rename operation.

userMetadata_username :: Lens' UserMetadata (Maybe Text)

The name of the user.

userMetadata_emailAddress :: Lens' UserMetadata (Maybe Text)

The email address of the user.

UserStorageMetadata

userStorageMetadata_storageRule :: Lens' UserStorageMetadata (Maybe StorageRuleType)

The storage for a user.

userStorageMetadata_storageUtilizedInBytes :: Lens' UserStorageMetadata (Maybe Integer)

The amount of storage used, in bytes.