Network.AWS.WorkDocs.RemoveResourcePermission

Description

Removes the permission for the specified principal from the specified resource.

Creating a Request

data RemoveResourcePermission

See: newRemoveResourcePermission smart constructor.

Constructors

RemoveResourcePermission'

Fields

principalType :: Maybe PrincipalType

The principal type of the resource.

authenticationToken :: Maybe (Sensitive Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

resourceId :: Text

The ID of the resource.

principalId :: Text

The principal ID of the resource.

newRemoveResourcePermission

Arguments

:: Text
$sel:resourceId:RemoveResourcePermission'

-> Text
$sel:principalId:RemoveResourcePermission'

-> RemoveResourcePermission

Create a value of RemoveResourcePermission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:principalType:RemoveResourcePermission', removeResourcePermission_principalType - The principal type of the resource.

$sel:authenticationToken:RemoveResourcePermission', removeResourcePermission_authenticationToken - Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

$sel:resourceId:RemoveResourcePermission', removeResourcePermission_resourceId - The ID of the resource.

$sel:principalId:RemoveResourcePermission', removeResourcePermission_principalId - The principal ID of the resource.

Request Lenses

removeResourcePermission_principalType :: Lens' RemoveResourcePermission (Maybe PrincipalType)

The principal type of the resource.

removeResourcePermission_authenticationToken :: Lens' RemoveResourcePermission (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

removeResourcePermission_resourceId :: Lens' RemoveResourcePermission Text

The ID of the resource.

removeResourcePermission_principalId :: Lens' RemoveResourcePermission Text

The principal ID of the resource.

Destructuring the Response

data RemoveResourcePermissionResponse

See: newRemoveResourcePermissionResponse smart constructor.

Constructors

RemoveResourcePermissionResponse'

newRemoveResourcePermissionResponse :: RemoveResourcePermissionResponse

Create a value of RemoveResourcePermissionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.