Network.AWS.WorkDocs.Types

Description

Service Configuration

defaultService :: Service

API version 2016-05-01 of the Amazon WorkDocs SDK configuration.

Errors

_CustomMetadataLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The limit has been reached on the number of custom properties for the specified resource.

_EntityAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The resource already exists.

_ResourceAlreadyCheckedOutException :: AsError a => Getting (First ServiceError) a ServiceError

The resource is already checked out.

_ProhibitedStateException :: AsError a => Getting (First ServiceError) a ServiceError

The specified document version is not in the INITIALIZED state.

_TooManyLabelsException :: AsError a => Getting (First ServiceError) a ServiceError

The limit has been reached on the number of labels for the specified resource.

_InvalidArgumentException :: AsError a => Getting (First ServiceError) a ServiceError

The pagination marker or limit fields are not valid.

_UnauthorizedResourceAccessException :: AsError a => Getting (First ServiceError) a ServiceError

The caller does not have access to perform the action on the resource.

_TooManySubscriptionsException :: AsError a => Getting (First ServiceError) a ServiceError

You've reached the limit on the number of subscriptions for the WorkDocs instance.

_FailedDependencyException :: AsError a => Getting (First ServiceError) a ServiceError

The AWS Directory Service cannot reach an on-premises instance. Or a dependency under the control of the organization is failing, such as a connected Active Directory.

_DocumentLockedForCommentsException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the document is locked for comments and user tries to create or delete a comment on that document.

_EntityNotExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The resource does not exist.

_ConflictingOperationException :: AsError a => Getting (First ServiceError) a ServiceError

Another operation is in progress on the resource that conflicts with the current operation.

_DeactivatingLastSystemUserException :: AsError a => Getting (First ServiceError) a ServiceError

The last user in the organization is being deactivated.

_IllegalUserStateException :: AsError a => Getting (First ServiceError) a ServiceError

The user is undergoing transfer of ownership.

_StorageLimitWillExceedException :: AsError a => Getting (First ServiceError) a ServiceError

The storage limit will be exceeded.

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

The resource hierarchy is changing.

_RequestedEntityTooLargeException :: AsError a => Getting (First ServiceError) a ServiceError

The response is too large to return. The request must include a filter to reduce the size of the response.

_StorageLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The storage limit has been exceeded.

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

One or more of the dependencies is unavailable.

_InvalidCommentOperationException :: AsError a => Getting (First ServiceError) a ServiceError

The requested operation is not allowed on the specified comment object.

_InvalidOperationException :: AsError a => Getting (First ServiceError) a ServiceError

The operation is invalid.

_UnauthorizedOperationException :: AsError a => Getting (First ServiceError) a ServiceError

The operation is not permitted.

_DraftUploadOutOfSyncException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when a valid checkout ID is not presented on document version upload calls for a document that has been checked out from Web client.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum of 100,000 folders under the parent folder has been exceeded.

_InvalidPasswordException :: AsError a => Getting (First ServiceError) a ServiceError

The password is invalid.

ActivityType

newtype ActivityType

Constructors

ActivityType'

Fields

fromActivityType :: Text

Bundled Patterns

pattern ActivityType_DOCUMENT_ANNOTATION_ADDED :: ActivityType

pattern ActivityType_DOCUMENT_ANNOTATION_DELETED :: ActivityType

pattern ActivityType_DOCUMENT_CHECKED_IN :: ActivityType

pattern ActivityType_DOCUMENT_CHECKED_OUT :: ActivityType

pattern ActivityType_DOCUMENT_COMMENT_ADDED :: ActivityType

pattern ActivityType_DOCUMENT_COMMENT_DELETED :: ActivityType

pattern ActivityType_DOCUMENT_MOVED :: ActivityType

pattern ActivityType_DOCUMENT_RECYCLED :: ActivityType

pattern ActivityType_DOCUMENT_RENAMED :: ActivityType

pattern ActivityType_DOCUMENT_RESTORED :: ActivityType

pattern ActivityType_DOCUMENT_REVERTED :: ActivityType

pattern ActivityType_DOCUMENT_SHAREABLE_LINK_CREATED :: ActivityType

pattern ActivityType_DOCUMENT_SHAREABLE_LINK_PERMISSION_CHANGED :: ActivityType

pattern ActivityType_DOCUMENT_SHAREABLE_LINK_REMOVED :: ActivityType

pattern ActivityType_DOCUMENT_SHARED :: ActivityType

pattern ActivityType_DOCUMENT_SHARE_PERMISSION_CHANGED :: ActivityType

pattern ActivityType_DOCUMENT_UNSHARED :: ActivityType

pattern ActivityType_DOCUMENT_VERSION_DELETED :: ActivityType

pattern ActivityType_DOCUMENT_VERSION_DOWNLOADED :: ActivityType

pattern ActivityType_DOCUMENT_VERSION_UPLOADED :: ActivityType

pattern ActivityType_DOCUMENT_VERSION_VIEWED :: ActivityType

pattern ActivityType_FOLDER_CREATED :: ActivityType

pattern ActivityType_FOLDER_DELETED :: ActivityType

pattern ActivityType_FOLDER_MOVED :: ActivityType

pattern ActivityType_FOLDER_RECYCLED :: ActivityType

pattern ActivityType_FOLDER_RENAMED :: ActivityType

pattern ActivityType_FOLDER_RESTORED :: ActivityType

pattern ActivityType_FOLDER_SHAREABLE_LINK_CREATED :: ActivityType

pattern ActivityType_FOLDER_SHAREABLE_LINK_PERMISSION_CHANGED :: ActivityType

pattern ActivityType_FOLDER_SHAREABLE_LINK_REMOVED :: ActivityType

pattern ActivityType_FOLDER_SHARED :: ActivityType

pattern ActivityType_FOLDER_SHARE_PERMISSION_CHANGED :: ActivityType

pattern ActivityType_FOLDER_UNSHARED :: ActivityType

BooleanEnumType

newtype BooleanEnumType

Constructors

BooleanEnumType'

Fields

fromBooleanEnumType :: Text

Bundled Patterns

pattern BooleanEnumType_FALSE :: BooleanEnumType

pattern BooleanEnumType_TRUE :: BooleanEnumType

CommentStatusType

newtype CommentStatusType

Constructors

CommentStatusType'

Fields

fromCommentStatusType :: Text

Bundled Patterns

pattern CommentStatusType_DELETED :: CommentStatusType

pattern CommentStatusType_DRAFT :: CommentStatusType

pattern CommentStatusType_PUBLISHED :: CommentStatusType

CommentVisibilityType

newtype CommentVisibilityType

Constructors

CommentVisibilityType'

Fields

fromCommentVisibilityType :: Text

Bundled Patterns

pattern CommentVisibilityType_PRIVATE :: CommentVisibilityType

pattern CommentVisibilityType_PUBLIC :: CommentVisibilityType

DocumentSourceType

newtype DocumentSourceType

Constructors

DocumentSourceType'

Fields

fromDocumentSourceType :: Text

Bundled Patterns

pattern DocumentSourceType_ORIGINAL :: DocumentSourceType

pattern DocumentSourceType_WITH_COMMENTS :: DocumentSourceType

DocumentStatusType

newtype DocumentStatusType

Constructors

DocumentStatusType'

Fields

fromDocumentStatusType :: Text

Bundled Patterns

pattern DocumentStatusType_ACTIVE :: DocumentStatusType

pattern DocumentStatusType_INITIALIZED :: DocumentStatusType

DocumentThumbnailType

newtype DocumentThumbnailType

Constructors

DocumentThumbnailType'

Fields

fromDocumentThumbnailType :: Text

Bundled Patterns

pattern DocumentThumbnailType_LARGE :: DocumentThumbnailType

pattern DocumentThumbnailType_SMALL :: DocumentThumbnailType

pattern DocumentThumbnailType_SMALL_HQ :: DocumentThumbnailType

DocumentVersionStatus

newtype DocumentVersionStatus

Constructors

DocumentVersionStatus'

Fields

fromDocumentVersionStatus :: Text

Bundled Patterns

pattern DocumentVersionStatus_ACTIVE :: DocumentVersionStatus

FolderContentType

newtype FolderContentType

Constructors

FolderContentType'

Fields

fromFolderContentType :: Text

Bundled Patterns

pattern FolderContentType_ALL :: FolderContentType

pattern FolderContentType_DOCUMENT :: FolderContentType

pattern FolderContentType_FOLDER :: FolderContentType

LocaleType

newtype LocaleType

Constructors

LocaleType'

Fields

fromLocaleType :: Text

Bundled Patterns

pattern LocaleType_De :: LocaleType

pattern LocaleType_Default :: LocaleType

pattern LocaleType_En :: LocaleType

pattern LocaleType_Es :: LocaleType

pattern LocaleType_Fr :: LocaleType

pattern LocaleType_Ja :: LocaleType

pattern LocaleType_Ko :: LocaleType

pattern LocaleType_Pt_BR :: LocaleType

pattern LocaleType_Ru :: LocaleType

pattern LocaleType_Zh_CN :: LocaleType

pattern LocaleType_Zh_TW :: LocaleType

OrderType

newtype OrderType

Constructors

OrderType'

Fields

fromOrderType :: Text

Bundled Patterns

pattern OrderType_ASCENDING :: OrderType

pattern OrderType_DESCENDING :: OrderType

PrincipalType

newtype PrincipalType

Constructors

PrincipalType'

Fields

fromPrincipalType :: Text

Bundled Patterns

pattern PrincipalType_ANONYMOUS :: PrincipalType

pattern PrincipalType_GROUP :: PrincipalType

pattern PrincipalType_INVITE :: PrincipalType

pattern PrincipalType_ORGANIZATION :: PrincipalType

pattern PrincipalType_USER :: PrincipalType

ResourceCollectionType

newtype ResourceCollectionType

Constructors

ResourceCollectionType'

Fields

fromResourceCollectionType :: Text

Bundled Patterns

pattern ResourceCollectionType_SHARED_WITH_ME :: ResourceCollectionType

ResourceSortType

newtype ResourceSortType

Constructors

ResourceSortType'

Fields

fromResourceSortType :: Text

Bundled Patterns

pattern ResourceSortType_DATE :: ResourceSortType

pattern ResourceSortType_NAME :: ResourceSortType

ResourceStateType

newtype ResourceStateType

Constructors

ResourceStateType'

Fields

fromResourceStateType :: Text

Bundled Patterns

pattern ResourceStateType_ACTIVE :: ResourceStateType

pattern ResourceStateType_RECYCLED :: ResourceStateType

pattern ResourceStateType_RECYCLING :: ResourceStateType

pattern ResourceStateType_RESTORING :: ResourceStateType

ResourceType

newtype ResourceType

Constructors

ResourceType'

Fields

fromResourceType :: Text

Bundled Patterns

pattern ResourceType_DOCUMENT :: ResourceType

pattern ResourceType_FOLDER :: ResourceType

RolePermissionType

newtype RolePermissionType

Constructors

RolePermissionType'

Fields

fromRolePermissionType :: Text

Bundled Patterns

pattern RolePermissionType_DIRECT :: RolePermissionType

pattern RolePermissionType_INHERITED :: RolePermissionType

RoleType

newtype RoleType

Constructors

RoleType'

Fields

fromRoleType :: Text

Bundled Patterns

pattern RoleType_CONTRIBUTOR :: RoleType

pattern RoleType_COOWNER :: RoleType

pattern RoleType_OWNER :: RoleType

pattern RoleType_VIEWER :: RoleType

ShareStatusType

newtype ShareStatusType

Constructors

ShareStatusType'

Fields

fromShareStatusType :: Text

Bundled Patterns

pattern ShareStatusType_FAILURE :: ShareStatusType

pattern ShareStatusType_SUCCESS :: ShareStatusType

StorageType

newtype StorageType

Constructors

StorageType'

Fields

fromStorageType :: Text

Bundled Patterns

pattern StorageType_QUOTA :: StorageType

pattern StorageType_UNLIMITED :: StorageType

SubscriptionProtocolType

newtype SubscriptionProtocolType

Constructors

SubscriptionProtocolType'

Fields

fromSubscriptionProtocolType :: Text

Bundled Patterns

pattern SubscriptionProtocolType_HTTPS :: SubscriptionProtocolType

SubscriptionType

newtype SubscriptionType

Constructors

SubscriptionType'

Fields

fromSubscriptionType :: Text

Bundled Patterns

pattern SubscriptionType_ALL :: SubscriptionType

UserFilterType

newtype UserFilterType

Constructors

UserFilterType'

Fields

fromUserFilterType :: Text

Bundled Patterns

pattern UserFilterType_ACTIVE_PENDING :: UserFilterType

pattern UserFilterType_ALL :: UserFilterType

UserSortType

newtype UserSortType

Constructors

UserSortType'

Fields

fromUserSortType :: Text

Bundled Patterns

pattern UserSortType_FULL_NAME :: UserSortType

pattern UserSortType_STORAGE_LIMIT :: UserSortType

pattern UserSortType_STORAGE_USED :: UserSortType

pattern UserSortType_USER_NAME :: UserSortType

pattern UserSortType_USER_STATUS :: UserSortType

UserStatusType

newtype UserStatusType

Constructors

UserStatusType'

Fields

fromUserStatusType :: Text

Bundled Patterns

pattern UserStatusType_ACTIVE :: UserStatusType

pattern UserStatusType_INACTIVE :: UserStatusType

pattern UserStatusType_PENDING :: UserStatusType

UserType

newtype UserType

Constructors

UserType'

Fields

fromUserType :: Text

Bundled Patterns

pattern UserType_ADMIN :: UserType

pattern UserType_MINIMALUSER :: UserType

pattern UserType_POWERUSER :: UserType

pattern UserType_USER :: UserType

pattern UserType_WORKSPACESUSER :: UserType

Activity

data Activity

Describes the activity information.

See: newActivity smart constructor.

Constructors

Activity'

Fields

resourceMetadata :: Maybe ResourceMetadata

The metadata of the resource involved in the user action.

isIndirectActivity :: Maybe Bool

Indicates whether an activity is indirect or direct. An indirect activity results from a direct activity performed on a parent resource. For example, sharing a parent folder (the direct activity) shares all of the subfolders and documents within the parent folder (the indirect activity).

initiator :: Maybe UserMetadata

The user who performed the action.

participants :: Maybe Participants

The list of users or groups impacted by this action. This is an optional field and is filled for the following sharing activities: DOCUMENT_SHARED, DOCUMENT_SHARED, DOCUMENT_UNSHARED, FOLDER_SHARED, FOLDER_UNSHARED.

originalParent :: Maybe ResourceMetadata

The original parent of the resource. This is an optional field and is filled for move activities.

type' :: Maybe ActivityType

The activity type.

commentMetadata :: Maybe CommentMetadata

Metadata of the commenting activity. This is an optional field and is filled for commenting activities.

timeStamp :: Maybe POSIX

The timestamp when the action was performed.

organizationId :: Maybe Text

The ID of the organization.

newActivity :: Activity

Create a value of Activity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceMetadata:Activity', activity_resourceMetadata - The metadata of the resource involved in the user action.

$sel:isIndirectActivity:Activity', activity_isIndirectActivity - Indicates whether an activity is indirect or direct. An indirect activity results from a direct activity performed on a parent resource. For example, sharing a parent folder (the direct activity) shares all of the subfolders and documents within the parent folder (the indirect activity).

$sel:initiator:Activity', activity_initiator - The user who performed the action.

$sel:participants:Activity', activity_participants - The list of users or groups impacted by this action. This is an optional field and is filled for the following sharing activities: DOCUMENT_SHARED, DOCUMENT_SHARED, DOCUMENT_UNSHARED, FOLDER_SHARED, FOLDER_UNSHARED.

$sel:originalParent:Activity', activity_originalParent - The original parent of the resource. This is an optional field and is filled for move activities.

$sel:type':Activity', activity_type - The activity type.

$sel:commentMetadata:Activity', activity_commentMetadata - Metadata of the commenting activity. This is an optional field and is filled for commenting activities.

$sel:timeStamp:Activity', activity_timeStamp - The timestamp when the action was performed.

$sel:organizationId:Activity', activity_organizationId - The ID of the organization.

activity_resourceMetadata :: Lens' Activity (Maybe ResourceMetadata)

The metadata of the resource involved in the user action.

activity_isIndirectActivity :: Lens' Activity (Maybe Bool)

Indicates whether an activity is indirect or direct. An indirect activity results from a direct activity performed on a parent resource. For example, sharing a parent folder (the direct activity) shares all of the subfolders and documents within the parent folder (the indirect activity).

activity_initiator :: Lens' Activity (Maybe UserMetadata)

The user who performed the action.

activity_participants :: Lens' Activity (Maybe Participants)

The list of users or groups impacted by this action. This is an optional field and is filled for the following sharing activities: DOCUMENT_SHARED, DOCUMENT_SHARED, DOCUMENT_UNSHARED, FOLDER_SHARED, FOLDER_UNSHARED.

activity_originalParent :: Lens' Activity (Maybe ResourceMetadata)

The original parent of the resource. This is an optional field and is filled for move activities.

activity_type :: Lens' Activity (Maybe ActivityType)

The activity type.

activity_commentMetadata :: Lens' Activity (Maybe CommentMetadata)

Metadata of the commenting activity. This is an optional field and is filled for commenting activities.

activity_timeStamp :: Lens' Activity (Maybe UTCTime)

The timestamp when the action was performed.

activity_organizationId :: Lens' Activity (Maybe Text)

The ID of the organization.

Comment

data Comment

Describes a comment.

See: newComment smart constructor.

Constructors

Comment'

Fields

status :: Maybe CommentStatusType

The status of the comment.

text :: Maybe (Sensitive Text)

The text of the comment.

visibility :: Maybe CommentVisibilityType

The visibility of the comment. Options are either PRIVATE, where the comment is visible only to the comment author and document owner and co-owners, or PUBLIC, where the comment is visible to document owners, co-owners, and contributors.

threadId :: Maybe Text

The ID of the root comment in the thread.

contributor :: Maybe User

The details of the user who made the comment.

createdTimestamp :: Maybe POSIX

The time that the comment was created.

recipientId :: Maybe Text

If the comment is a reply to another user's comment, this field contains the user ID of the user being replied to.

parentId :: Maybe Text

The ID of the parent comment.

commentId :: Text

The ID of the comment.

newComment

Arguments

:: Text
$sel:commentId:Comment'

-> Comment

Create a value of Comment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Comment', comment_status - The status of the comment.

$sel:text:Comment', comment_text - The text of the comment.

$sel:visibility:Comment', comment_visibility - The visibility of the comment. Options are either PRIVATE, where the comment is visible only to the comment author and document owner and co-owners, or PUBLIC, where the comment is visible to document owners, co-owners, and contributors.

$sel:threadId:Comment', comment_threadId - The ID of the root comment in the thread.

$sel:contributor:Comment', comment_contributor - The details of the user who made the comment.

$sel:createdTimestamp:Comment', comment_createdTimestamp - The time that the comment was created.

$sel:recipientId:Comment', comment_recipientId - If the comment is a reply to another user's comment, this field contains the user ID of the user being replied to.

$sel:parentId:Comment', comment_parentId - The ID of the parent comment.

$sel:commentId:Comment', comment_commentId - The ID of the comment.

comment_status :: Lens' Comment (Maybe CommentStatusType)

The status of the comment.

comment_text :: Lens' Comment (Maybe Text)

The text of the comment.

comment_visibility :: Lens' Comment (Maybe CommentVisibilityType)

The visibility of the comment. Options are either PRIVATE, where the comment is visible only to the comment author and document owner and co-owners, or PUBLIC, where the comment is visible to document owners, co-owners, and contributors.

comment_threadId :: Lens' Comment (Maybe Text)

The ID of the root comment in the thread.

comment_contributor :: Lens' Comment (Maybe User)

The details of the user who made the comment.

comment_createdTimestamp :: Lens' Comment (Maybe UTCTime)

The time that the comment was created.

comment_recipientId :: Lens' Comment (Maybe Text)

If the comment is a reply to another user's comment, this field contains the user ID of the user being replied to.

comment_parentId :: Lens' Comment (Maybe Text)

The ID of the parent comment.

comment_commentId :: Lens' Comment Text

The ID of the comment.

CommentMetadata

data CommentMetadata

Describes the metadata of a comment.

See: newCommentMetadata smart constructor.

Constructors

CommentMetadata'

Fields

commentStatus :: Maybe CommentStatusType

The status of the comment.

contributor :: Maybe User

The user who made the comment.

commentId :: Maybe Text

The ID of the comment.

createdTimestamp :: Maybe POSIX

The timestamp that the comment was created.

recipientId :: Maybe Text

The ID of the user being replied to.

newCommentMetadata :: CommentMetadata

Create a value of CommentMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:commentStatus:CommentMetadata', commentMetadata_commentStatus - The status of the comment.

$sel:contributor:CommentMetadata', commentMetadata_contributor - The user who made the comment.

$sel:commentId:CommentMetadata', commentMetadata_commentId - The ID of the comment.

$sel:createdTimestamp:CommentMetadata', commentMetadata_createdTimestamp - The timestamp that the comment was created.

$sel:recipientId:CommentMetadata', commentMetadata_recipientId - The ID of the user being replied to.

commentMetadata_commentStatus :: Lens' CommentMetadata (Maybe CommentStatusType)

The status of the comment.

commentMetadata_contributor :: Lens' CommentMetadata (Maybe User)

The user who made the comment.

commentMetadata_commentId :: Lens' CommentMetadata (Maybe Text)

The ID of the comment.

commentMetadata_createdTimestamp :: Lens' CommentMetadata (Maybe UTCTime)

The timestamp that the comment was created.

commentMetadata_recipientId :: Lens' CommentMetadata (Maybe Text)

The ID of the user being replied to.

DocumentMetadata

data DocumentMetadata

Describes the document.

See: newDocumentMetadata smart constructor.

Constructors

DocumentMetadata'

Fields

latestVersionMetadata :: Maybe DocumentVersionMetadata

The latest version of the document.

parentFolderId :: Maybe Text

The ID of the parent folder.

modifiedTimestamp :: Maybe POSIX

The time when the document was updated.

id :: Maybe Text

The ID of the document.

labels :: Maybe [Text]

List of labels on the document.

resourceState :: Maybe ResourceStateType

The resource state.

createdTimestamp :: Maybe POSIX

The time when the document was created.

creatorId :: Maybe Text

The ID of the creator.

newDocumentMetadata :: DocumentMetadata

Create a value of DocumentMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:latestVersionMetadata:DocumentMetadata', documentMetadata_latestVersionMetadata - The latest version of the document.

$sel:parentFolderId:DocumentMetadata', documentMetadata_parentFolderId - The ID of the parent folder.

$sel:modifiedTimestamp:DocumentMetadata', documentMetadata_modifiedTimestamp - The time when the document was updated.

$sel:id:DocumentMetadata', documentMetadata_id - The ID of the document.

$sel:labels:DocumentMetadata', documentMetadata_labels - List of labels on the document.

$sel:resourceState:DocumentMetadata', documentMetadata_resourceState - The resource state.

$sel:createdTimestamp:DocumentMetadata', documentMetadata_createdTimestamp - The time when the document was created.

$sel:creatorId:DocumentMetadata', documentMetadata_creatorId - The ID of the creator.

documentMetadata_latestVersionMetadata :: Lens' DocumentMetadata (Maybe DocumentVersionMetadata)

The latest version of the document.

documentMetadata_parentFolderId :: Lens' DocumentMetadata (Maybe Text)

The ID of the parent folder.

documentMetadata_modifiedTimestamp :: Lens' DocumentMetadata (Maybe UTCTime)

The time when the document was updated.

documentMetadata_id :: Lens' DocumentMetadata (Maybe Text)

The ID of the document.

documentMetadata_labels :: Lens' DocumentMetadata (Maybe [Text])

List of labels on the document.

documentMetadata_resourceState :: Lens' DocumentMetadata (Maybe ResourceStateType)

The resource state.

documentMetadata_createdTimestamp :: Lens' DocumentMetadata (Maybe UTCTime)

The time when the document was created.

documentMetadata_creatorId :: Lens' DocumentMetadata (Maybe Text)

The ID of the creator.

DocumentVersionMetadata

data DocumentVersionMetadata

Describes a version of a document.

See: newDocumentVersionMetadata smart constructor.

Constructors

DocumentVersionMetadata'

Fields

thumbnail :: Maybe (HashMap DocumentThumbnailType (Sensitive Text))

The thumbnail of the document.

status :: Maybe DocumentStatusType

The status of the document.

signature :: Maybe Text

The signature of the document.

contentCreatedTimestamp :: Maybe POSIX

The timestamp when the content of the document was originally created.

size :: Maybe Integer

The size of the document, in bytes.

name :: Maybe Text

The name of the version.

modifiedTimestamp :: Maybe POSIX

The timestamp when the document was last uploaded.

source :: Maybe (HashMap DocumentSourceType (Sensitive Text))

The source of the document.

id :: Maybe Text

The ID of the version.

createdTimestamp :: Maybe POSIX

The timestamp when the document was first uploaded.

contentModifiedTimestamp :: Maybe POSIX

The timestamp when the content of the document was modified.

creatorId :: Maybe Text

The ID of the creator.

contentType :: Maybe Text

The content type of the document.

newDocumentVersionMetadata :: DocumentVersionMetadata

Create a value of DocumentVersionMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thumbnail:DocumentVersionMetadata', documentVersionMetadata_thumbnail - The thumbnail of the document.

$sel:status:DocumentVersionMetadata', documentVersionMetadata_status - The status of the document.

$sel:signature:DocumentVersionMetadata', documentVersionMetadata_signature - The signature of the document.

$sel:contentCreatedTimestamp:DocumentVersionMetadata', documentVersionMetadata_contentCreatedTimestamp - The timestamp when the content of the document was originally created.

$sel:size:DocumentVersionMetadata', documentVersionMetadata_size - The size of the document, in bytes.

$sel:name:DocumentVersionMetadata', documentVersionMetadata_name - The name of the version.

$sel:modifiedTimestamp:DocumentVersionMetadata', documentVersionMetadata_modifiedTimestamp - The timestamp when the document was last uploaded.

$sel:source:DocumentVersionMetadata', documentVersionMetadata_source - The source of the document.

$sel:id:DocumentVersionMetadata', documentVersionMetadata_id - The ID of the version.

$sel:createdTimestamp:DocumentVersionMetadata', documentVersionMetadata_createdTimestamp - The timestamp when the document was first uploaded.

$sel:contentModifiedTimestamp:DocumentVersionMetadata', documentVersionMetadata_contentModifiedTimestamp - The timestamp when the content of the document was modified.

$sel:creatorId:DocumentVersionMetadata', documentVersionMetadata_creatorId - The ID of the creator.

$sel:contentType:DocumentVersionMetadata', documentVersionMetadata_contentType - The content type of the document.

documentVersionMetadata_thumbnail :: Lens' DocumentVersionMetadata (Maybe (HashMap DocumentThumbnailType Text))

The thumbnail of the document.

documentVersionMetadata_status :: Lens' DocumentVersionMetadata (Maybe DocumentStatusType)

The status of the document.

documentVersionMetadata_signature :: Lens' DocumentVersionMetadata (Maybe Text)

The signature of the document.

documentVersionMetadata_contentCreatedTimestamp :: Lens' DocumentVersionMetadata (Maybe UTCTime)

The timestamp when the content of the document was originally created.

documentVersionMetadata_size :: Lens' DocumentVersionMetadata (Maybe Integer)

The size of the document, in bytes.

documentVersionMetadata_name :: Lens' DocumentVersionMetadata (Maybe Text)

The name of the version.

documentVersionMetadata_modifiedTimestamp :: Lens' DocumentVersionMetadata (Maybe UTCTime)

The timestamp when the document was last uploaded.

documentVersionMetadata_source :: Lens' DocumentVersionMetadata (Maybe (HashMap DocumentSourceType Text))

The source of the document.

documentVersionMetadata_id :: Lens' DocumentVersionMetadata (Maybe Text)

The ID of the version.

documentVersionMetadata_createdTimestamp :: Lens' DocumentVersionMetadata (Maybe UTCTime)

The timestamp when the document was first uploaded.

documentVersionMetadata_contentModifiedTimestamp :: Lens' DocumentVersionMetadata (Maybe UTCTime)

The timestamp when the content of the document was modified.

documentVersionMetadata_creatorId :: Lens' DocumentVersionMetadata (Maybe Text)

The ID of the creator.

documentVersionMetadata_contentType :: Lens' DocumentVersionMetadata (Maybe Text)

The content type of the document.

FolderMetadata

data FolderMetadata

Describes a folder.

See: newFolderMetadata smart constructor.

Constructors

FolderMetadata'

Fields

signature :: Maybe Text

The unique identifier created from the subfolders and documents of the folder.

parentFolderId :: Maybe Text

The ID of the parent folder.

size :: Maybe Integer

The size of the folder metadata.

latestVersionSize :: Maybe Integer

The size of the latest version of the folder metadata.

name :: Maybe Text

The name of the folder.

modifiedTimestamp :: Maybe POSIX

The time when the folder was updated.

id :: Maybe Text

The ID of the folder.

labels :: Maybe [Text]

List of labels on the folder.

resourceState :: Maybe ResourceStateType

The resource state of the folder.

createdTimestamp :: Maybe POSIX

The time when the folder was created.

creatorId :: Maybe Text

The ID of the creator.

newFolderMetadata :: FolderMetadata

Create a value of FolderMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:signature:FolderMetadata', folderMetadata_signature - The unique identifier created from the subfolders and documents of the folder.

$sel:parentFolderId:FolderMetadata', folderMetadata_parentFolderId - The ID of the parent folder.

$sel:size:FolderMetadata', folderMetadata_size - The size of the folder metadata.

$sel:latestVersionSize:FolderMetadata', folderMetadata_latestVersionSize - The size of the latest version of the folder metadata.

$sel:name:FolderMetadata', folderMetadata_name - The name of the folder.

$sel:modifiedTimestamp:FolderMetadata', folderMetadata_modifiedTimestamp - The time when the folder was updated.

$sel:id:FolderMetadata', folderMetadata_id - The ID of the folder.

$sel:labels:FolderMetadata', folderMetadata_labels - List of labels on the folder.

$sel:resourceState:FolderMetadata', folderMetadata_resourceState - The resource state of the folder.

$sel:createdTimestamp:FolderMetadata', folderMetadata_createdTimestamp - The time when the folder was created.

$sel:creatorId:FolderMetadata', folderMetadata_creatorId - The ID of the creator.

folderMetadata_signature :: Lens' FolderMetadata (Maybe Text)

The unique identifier created from the subfolders and documents of the folder.

folderMetadata_parentFolderId :: Lens' FolderMetadata (Maybe Text)

The ID of the parent folder.

folderMetadata_size :: Lens' FolderMetadata (Maybe Integer)

The size of the folder metadata.

folderMetadata_latestVersionSize :: Lens' FolderMetadata (Maybe Integer)

The size of the latest version of the folder metadata.

folderMetadata_name :: Lens' FolderMetadata (Maybe Text)

The name of the folder.

folderMetadata_modifiedTimestamp :: Lens' FolderMetadata (Maybe UTCTime)

The time when the folder was updated.

folderMetadata_id :: Lens' FolderMetadata (Maybe Text)

The ID of the folder.

folderMetadata_labels :: Lens' FolderMetadata (Maybe [Text])

List of labels on the folder.

folderMetadata_resourceState :: Lens' FolderMetadata (Maybe ResourceStateType)

The resource state of the folder.

folderMetadata_createdTimestamp :: Lens' FolderMetadata (Maybe UTCTime)

The time when the folder was created.

folderMetadata_creatorId :: Lens' FolderMetadata (Maybe Text)

The ID of the creator.

GroupMetadata

data GroupMetadata

Describes the metadata of a user group.

See: newGroupMetadata smart constructor.

Constructors

GroupMetadata'

Fields

name :: Maybe Text

The name of the group.

id :: Maybe Text

The ID of the user group.

newGroupMetadata :: GroupMetadata

Create a value of GroupMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GroupMetadata', groupMetadata_name - The name of the group.

$sel:id:GroupMetadata', groupMetadata_id - The ID of the user group.

groupMetadata_name :: Lens' GroupMetadata (Maybe Text)

The name of the group.

groupMetadata_id :: Lens' GroupMetadata (Maybe Text)

The ID of the user group.

NotificationOptions

data NotificationOptions

Set of options which defines notification preferences of given action.

See: newNotificationOptions smart constructor.

Constructors

NotificationOptions'

Fields

emailMessage :: Maybe (Sensitive Text)

Text value to be included in the email body.

sendEmail :: Maybe Bool

Boolean value to indicate an email notification should be sent to the receipients.

newNotificationOptions :: NotificationOptions

Create a value of NotificationOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:emailMessage:NotificationOptions', notificationOptions_emailMessage - Text value to be included in the email body.

$sel:sendEmail:NotificationOptions', notificationOptions_sendEmail - Boolean value to indicate an email notification should be sent to the receipients.

notificationOptions_emailMessage :: Lens' NotificationOptions (Maybe Text)

Text value to be included in the email body.

notificationOptions_sendEmail :: Lens' NotificationOptions (Maybe Bool)

Boolean value to indicate an email notification should be sent to the receipients.

Participants

data Participants

Describes the users or user groups.

See: newParticipants smart constructor.

Constructors

Participants'

Fields

groups :: Maybe [GroupMetadata]

The list of user groups.

users :: Maybe [UserMetadata]

The list of users.

newParticipants :: Participants

Create a value of Participants with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groups:Participants', participants_groups - The list of user groups.

$sel:users:Participants', participants_users - The list of users.

participants_groups :: Lens' Participants (Maybe [GroupMetadata])

The list of user groups.

participants_users :: Lens' Participants (Maybe [UserMetadata])

The list of users.

PermissionInfo

data PermissionInfo

Describes the permissions.

See: newPermissionInfo smart constructor.

Constructors

PermissionInfo'

Fields

role' :: Maybe RoleType

The role of the user.

type' :: Maybe RolePermissionType

The type of permissions.

newPermissionInfo :: PermissionInfo

Create a value of PermissionInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:role':PermissionInfo', permissionInfo_role - The role of the user.

$sel:type':PermissionInfo', permissionInfo_type - The type of permissions.

permissionInfo_role :: Lens' PermissionInfo (Maybe RoleType)

The role of the user.

permissionInfo_type :: Lens' PermissionInfo (Maybe RolePermissionType)

The type of permissions.

Principal

data Principal

Describes a resource.

See: newPrincipal smart constructor.

Constructors

Principal'

Fields

roles :: Maybe [PermissionInfo]

The permission information for the resource.

id :: Maybe Text

The ID of the resource.

type' :: Maybe PrincipalType

The type of resource.

newPrincipal :: Principal

Create a value of Principal with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roles:Principal', principal_roles - The permission information for the resource.

$sel:id:Principal', principal_id - The ID of the resource.

$sel:type':Principal', principal_type - The type of resource.

principal_roles :: Lens' Principal (Maybe [PermissionInfo])

The permission information for the resource.

principal_id :: Lens' Principal (Maybe Text)

The ID of the resource.

principal_type :: Lens' Principal (Maybe PrincipalType)

The type of resource.

ResourceMetadata

data ResourceMetadata

Describes the metadata of a resource.

See: newResourceMetadata smart constructor.

Constructors

ResourceMetadata'

Fields

versionId :: Maybe Text

The version ID of the resource. This is an optional field and is filled for action on document version.

owner :: Maybe UserMetadata

The owner of the resource.

name :: Maybe Text

The name of the resource.

id :: Maybe Text

The ID of the resource.

type' :: Maybe ResourceType

The type of resource.

originalName :: Maybe Text

The original name of the resource before a rename operation.

parentId :: Maybe Text

The parent ID of the resource before a rename operation.

newResourceMetadata :: ResourceMetadata

Create a value of ResourceMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:ResourceMetadata', resourceMetadata_versionId - The version ID of the resource. This is an optional field and is filled for action on document version.

$sel:owner:ResourceMetadata', resourceMetadata_owner - The owner of the resource.

$sel:name:ResourceMetadata', resourceMetadata_name - The name of the resource.

$sel:id:ResourceMetadata', resourceMetadata_id - The ID of the resource.

$sel:type':ResourceMetadata', resourceMetadata_type - The type of resource.

$sel:originalName:ResourceMetadata', resourceMetadata_originalName - The original name of the resource before a rename operation.

$sel:parentId:ResourceMetadata', resourceMetadata_parentId - The parent ID of the resource before a rename operation.

resourceMetadata_versionId :: Lens' ResourceMetadata (Maybe Text)

The version ID of the resource. This is an optional field and is filled for action on document version.

resourceMetadata_owner :: Lens' ResourceMetadata (Maybe UserMetadata)

The owner of the resource.

resourceMetadata_name :: Lens' ResourceMetadata (Maybe Text)

The name of the resource.

resourceMetadata_id :: Lens' ResourceMetadata (Maybe Text)

The ID of the resource.

resourceMetadata_type :: Lens' ResourceMetadata (Maybe ResourceType)

The type of resource.

resourceMetadata_originalName :: Lens' ResourceMetadata (Maybe Text)

The original name of the resource before a rename operation.

resourceMetadata_parentId :: Lens' ResourceMetadata (Maybe Text)

The parent ID of the resource before a rename operation.

ResourcePath

data ResourcePath

Describes the path information of a resource.

See: newResourcePath smart constructor.

Constructors

ResourcePath'

Fields

components :: Maybe [ResourcePathComponent]

The components of the resource path.

newResourcePath :: ResourcePath

Create a value of ResourcePath with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:components:ResourcePath', resourcePath_components - The components of the resource path.

resourcePath_components :: Lens' ResourcePath (Maybe [ResourcePathComponent])

The components of the resource path.

ResourcePathComponent

data ResourcePathComponent

Describes the resource path.

See: newResourcePathComponent smart constructor.

Constructors

ResourcePathComponent'

Fields

name :: Maybe Text

The name of the resource path.

id :: Maybe Text

The ID of the resource path.

newResourcePathComponent :: ResourcePathComponent

Create a value of ResourcePathComponent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ResourcePathComponent', resourcePathComponent_name - The name of the resource path.

$sel:id:ResourcePathComponent', resourcePathComponent_id - The ID of the resource path.

resourcePathComponent_name :: Lens' ResourcePathComponent (Maybe Text)

The name of the resource path.

resourcePathComponent_id :: Lens' ResourcePathComponent (Maybe Text)

The ID of the resource path.

SharePrincipal

data SharePrincipal

Describes the recipient type and ID, if available.

See: newSharePrincipal smart constructor.

Constructors

SharePrincipal'

Fields

id :: Text

The ID of the recipient.

type' :: PrincipalType

The type of the recipient.

role' :: RoleType

The role of the recipient.

newSharePrincipal

Arguments

:: Text
$sel:id:SharePrincipal'

-> PrincipalType
$sel:type':SharePrincipal'

-> RoleType
$sel:role':SharePrincipal'

-> SharePrincipal

Create a value of SharePrincipal with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:SharePrincipal', sharePrincipal_id - The ID of the recipient.

$sel:type':SharePrincipal', sharePrincipal_type - The type of the recipient.

$sel:role':SharePrincipal', sharePrincipal_role - The role of the recipient.

sharePrincipal_id :: Lens' SharePrincipal Text

The ID of the recipient.

sharePrincipal_type :: Lens' SharePrincipal PrincipalType

The type of the recipient.

sharePrincipal_role :: Lens' SharePrincipal RoleType

The role of the recipient.

ShareResult

data ShareResult

Describes the share results of a resource.

See: newShareResult smart constructor.

Constructors

ShareResult'

Fields

status :: Maybe ShareStatusType

The status.

principalId :: Maybe Text

The ID of the principal.

inviteePrincipalId :: Maybe Text

The ID of the invited user.

role' :: Maybe RoleType

The role.

statusMessage :: Maybe (Sensitive Text)

The status message.

shareId :: Maybe Text

The ID of the resource that was shared.

newShareResult :: ShareResult

Create a value of ShareResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ShareResult', shareResult_status - The status.

$sel:principalId:ShareResult', shareResult_principalId - The ID of the principal.

$sel:inviteePrincipalId:ShareResult', shareResult_inviteePrincipalId - The ID of the invited user.

$sel:role':ShareResult', shareResult_role - The role.

$sel:statusMessage:ShareResult', shareResult_statusMessage - The status message.

$sel:shareId:ShareResult', shareResult_shareId - The ID of the resource that was shared.

shareResult_status :: Lens' ShareResult (Maybe ShareStatusType)

The status.

shareResult_principalId :: Lens' ShareResult (Maybe Text)

The ID of the principal.

shareResult_inviteePrincipalId :: Lens' ShareResult (Maybe Text)

The ID of the invited user.

shareResult_role :: Lens' ShareResult (Maybe RoleType)

The role.

shareResult_statusMessage :: Lens' ShareResult (Maybe Text)

The status message.

shareResult_shareId :: Lens' ShareResult (Maybe Text)

The ID of the resource that was shared.

StorageRuleType

data StorageRuleType

Describes the storage for a user.

See: newStorageRuleType smart constructor.

Constructors

StorageRuleType'

Fields

storageAllocatedInBytes :: Maybe Natural

The amount of storage allocated, in bytes.

storageType :: Maybe StorageType

The type of storage.

newStorageRuleType :: StorageRuleType

Create a value of StorageRuleType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:storageAllocatedInBytes:StorageRuleType', storageRuleType_storageAllocatedInBytes - The amount of storage allocated, in bytes.

$sel:storageType:StorageRuleType', storageRuleType_storageType - The type of storage.

storageRuleType_storageAllocatedInBytes :: Lens' StorageRuleType (Maybe Natural)

The amount of storage allocated, in bytes.

storageRuleType_storageType :: Lens' StorageRuleType (Maybe StorageType)

The type of storage.

Subscription

data Subscription

Describes a subscription.

See: newSubscription smart constructor.

Constructors

Subscription'

Fields

protocol :: Maybe SubscriptionProtocolType

The protocol of the subscription.

endPoint :: Maybe Text

The endpoint of the subscription.

subscriptionId :: Maybe Text

The ID of the subscription.

newSubscription :: Subscription

Create a value of Subscription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocol:Subscription', subscription_protocol - The protocol of the subscription.

$sel:endPoint:Subscription', subscription_endPoint - The endpoint of the subscription.

$sel:subscriptionId:Subscription', subscription_subscriptionId - The ID of the subscription.

subscription_protocol :: Lens' Subscription (Maybe SubscriptionProtocolType)

The protocol of the subscription.

subscription_endPoint :: Lens' Subscription (Maybe Text)

The endpoint of the subscription.

subscription_subscriptionId :: Lens' Subscription (Maybe Text)

The ID of the subscription.

UploadMetadata

data UploadMetadata

Describes the upload.

See: newUploadMetadata smart constructor.

Constructors

UploadMetadata'

Fields

uploadUrl :: Maybe (Sensitive Text)

The URL of the upload.

signedHeaders :: Maybe (HashMap Text Text)

The signed headers.

newUploadMetadata :: UploadMetadata

Create a value of UploadMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:uploadUrl:UploadMetadata', uploadMetadata_uploadUrl - The URL of the upload.

$sel:signedHeaders:UploadMetadata', uploadMetadata_signedHeaders - The signed headers.

uploadMetadata_uploadUrl :: Lens' UploadMetadata (Maybe Text)

The URL of the upload.

uploadMetadata_signedHeaders :: Lens' UploadMetadata (Maybe (HashMap Text Text))

The signed headers.

User

data User

Describes a user.

See: newUser smart constructor.

Constructors

User'

Fields

givenName :: Maybe Text

The given name of the user.

status :: Maybe UserStatusType

The status of the user.

locale :: Maybe LocaleType

The locale of the user.

username :: Maybe Text

The login name of the user.

storage :: Maybe UserStorageMetadata

The storage for the user.

modifiedTimestamp :: Maybe POSIX

The time when the user was modified.

emailAddress :: Maybe Text

The email address of the user.

id :: Maybe Text

The ID of the user.

rootFolderId :: Maybe Text

The ID of the root folder.

type' :: Maybe UserType

The type of user.

surname :: Maybe Text

The surname of the user.

timeZoneId :: Maybe Text

The time zone ID of the user.

createdTimestamp :: Maybe POSIX

The time when the user was created.

organizationId :: Maybe Text

The ID of the organization.

recycleBinFolderId :: Maybe Text

The ID of the recycle bin folder.

newUser :: User

Create a value of User with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:givenName:User', user_givenName - The given name of the user.

$sel:status:User', user_status - The status of the user.

$sel:locale:User', user_locale - The locale of the user.

$sel:username:User', user_username - The login name of the user.

$sel:storage:User', user_storage - The storage for the user.

$sel:modifiedTimestamp:User', user_modifiedTimestamp - The time when the user was modified.

$sel:emailAddress:User', user_emailAddress - The email address of the user.

$sel:id:User', user_id - The ID of the user.

$sel:rootFolderId:User', user_rootFolderId - The ID of the root folder.

$sel:type':User', user_type - The type of user.

$sel:surname:User', user_surname - The surname of the user.

$sel:timeZoneId:User', user_timeZoneId - The time zone ID of the user.

$sel:createdTimestamp:User', user_createdTimestamp - The time when the user was created.

$sel:organizationId:User', user_organizationId - The ID of the organization.

$sel:recycleBinFolderId:User', user_recycleBinFolderId - The ID of the recycle bin folder.

user_givenName :: Lens' User (Maybe Text)

The given name of the user.

user_status :: Lens' User (Maybe UserStatusType)

The status of the user.

user_locale :: Lens' User (Maybe LocaleType)

The locale of the user.

user_username :: Lens' User (Maybe Text)

The login name of the user.

user_storage :: Lens' User (Maybe UserStorageMetadata)

The storage for the user.

user_modifiedTimestamp :: Lens' User (Maybe UTCTime)

The time when the user was modified.

user_emailAddress :: Lens' User (Maybe Text)

The email address of the user.

user_id :: Lens' User (Maybe Text)

The ID of the user.

user_rootFolderId :: Lens' User (Maybe Text)

The ID of the root folder.

user_type :: Lens' User (Maybe UserType)

The type of user.

user_surname :: Lens' User (Maybe Text)

The surname of the user.

user_timeZoneId :: Lens' User (Maybe Text)

The time zone ID of the user.

user_createdTimestamp :: Lens' User (Maybe UTCTime)

The time when the user was created.

user_organizationId :: Lens' User (Maybe Text)

The ID of the organization.

user_recycleBinFolderId :: Lens' User (Maybe Text)

The ID of the recycle bin folder.

UserMetadata

data UserMetadata

Describes the metadata of the user.

See: newUserMetadata smart constructor.

Constructors

UserMetadata'

Fields

givenName :: Maybe Text

The given name of the user before a rename operation.

username :: Maybe Text

The name of the user.

emailAddress :: Maybe Text

The email address of the user.

id :: Maybe Text

The ID of the user.

surname :: Maybe Text

The surname of the user.

newUserMetadata :: UserMetadata

Create a value of UserMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:givenName:UserMetadata', userMetadata_givenName - The given name of the user before a rename operation.

$sel:username:UserMetadata', userMetadata_username - The name of the user.

$sel:emailAddress:UserMetadata', userMetadata_emailAddress - The email address of the user.

$sel:id:UserMetadata', userMetadata_id - The ID of the user.

$sel:surname:UserMetadata', userMetadata_surname - The surname of the user.

userMetadata_givenName :: Lens' UserMetadata (Maybe Text)

The given name of the user before a rename operation.

userMetadata_username :: Lens' UserMetadata (Maybe Text)

The name of the user.

userMetadata_emailAddress :: Lens' UserMetadata (Maybe Text)

The email address of the user.

userMetadata_id :: Lens' UserMetadata (Maybe Text)

The ID of the user.

userMetadata_surname :: Lens' UserMetadata (Maybe Text)

The surname of the user.

UserStorageMetadata

data UserStorageMetadata

Describes the storage for a user.

See: newUserStorageMetadata smart constructor.

Constructors

UserStorageMetadata'

Fields

storageUtilizedInBytes :: Maybe Integer

The amount of storage used, in bytes.

storageRule :: Maybe StorageRuleType

The storage for a user.

newUserStorageMetadata :: UserStorageMetadata

Create a value of UserStorageMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:storageUtilizedInBytes:UserStorageMetadata', userStorageMetadata_storageUtilizedInBytes - The amount of storage used, in bytes.

$sel:storageRule:UserStorageMetadata', userStorageMetadata_storageRule - The storage for a user.

userStorageMetadata_storageUtilizedInBytes :: Lens' UserStorageMetadata (Maybe Integer)

The amount of storage used, in bytes.

userStorageMetadata_storageRule :: Lens' UserStorageMetadata (Maybe StorageRuleType)

The storage for a user.