Network.AWS.WorkDocs.Types.Activity

Description

Documentation

data Activity

Describes the activity information.

See: newActivity smart constructor.

Constructors

Activity'

Fields

resourceMetadata :: Maybe ResourceMetadata

The metadata of the resource involved in the user action.

organizationId :: Maybe Text

The ID of the organization.

originalParent :: Maybe ResourceMetadata

The original parent of the resource. This is an optional field and is filled for move activities.

participants :: Maybe Participants

The list of users or groups impacted by this action. This is an optional field and is filled for the following sharing activities: DOCUMENT_SHARED, DOCUMENT_SHARED, DOCUMENT_UNSHARED, FOLDER_SHARED, FOLDER_UNSHARED.

commentMetadata :: Maybe CommentMetadata

Metadata of the commenting activity. This is an optional field and is filled for commenting activities.

timeStamp :: Maybe POSIX

The timestamp when the action was performed.

initiator :: Maybe UserMetadata

The user who performed the action.

type' :: Maybe ActivityType

The activity type.

isIndirectActivity :: Maybe Bool

Indicates whether an activity is indirect or direct. An indirect activity results from a direct activity performed on a parent resource. For example, sharing a parent folder (the direct activity) shares all of the subfolders and documents within the parent folder (the indirect activity).

newActivity :: Activity

Create a value of Activity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceMetadata:Activity', activity_resourceMetadata - The metadata of the resource involved in the user action.

$sel:organizationId:Activity', activity_organizationId - The ID of the organization.

$sel:originalParent:Activity', activity_originalParent - The original parent of the resource. This is an optional field and is filled for move activities.

$sel:participants:Activity', activity_participants - The list of users or groups impacted by this action. This is an optional field and is filled for the following sharing activities: DOCUMENT_SHARED, DOCUMENT_SHARED, DOCUMENT_UNSHARED, FOLDER_SHARED, FOLDER_UNSHARED.

$sel:commentMetadata:Activity', activity_commentMetadata - Metadata of the commenting activity. This is an optional field and is filled for commenting activities.

$sel:timeStamp:Activity', activity_timeStamp - The timestamp when the action was performed.

$sel:initiator:Activity', activity_initiator - The user who performed the action.

$sel:type':Activity', activity_type - The activity type.

$sel:isIndirectActivity:Activity', activity_isIndirectActivity - Indicates whether an activity is indirect or direct. An indirect activity results from a direct activity performed on a parent resource. For example, sharing a parent folder (the direct activity) shares all of the subfolders and documents within the parent folder (the indirect activity).

activity_resourceMetadata :: Lens' Activity (Maybe ResourceMetadata)

The metadata of the resource involved in the user action.

activity_organizationId :: Lens' Activity (Maybe Text)

The ID of the organization.

activity_originalParent :: Lens' Activity (Maybe ResourceMetadata)

The original parent of the resource. This is an optional field and is filled for move activities.

activity_participants :: Lens' Activity (Maybe Participants)

The list of users or groups impacted by this action. This is an optional field and is filled for the following sharing activities: DOCUMENT_SHARED, DOCUMENT_SHARED, DOCUMENT_UNSHARED, FOLDER_SHARED, FOLDER_UNSHARED.

activity_commentMetadata :: Lens' Activity (Maybe CommentMetadata)

Metadata of the commenting activity. This is an optional field and is filled for commenting activities.

activity_timeStamp :: Lens' Activity (Maybe UTCTime)

The timestamp when the action was performed.

activity_initiator :: Lens' Activity (Maybe UserMetadata)

The user who performed the action.

activity_type :: Lens' Activity (Maybe ActivityType)

The activity type.

activity_isIndirectActivity :: Lens' Activity (Maybe Bool)

Indicates whether an activity is indirect or direct. An indirect activity results from a direct activity performed on a parent resource. For example, sharing a parent folder (the direct activity) shares all of the subfolders and documents within the parent folder (the indirect activity).