Network.AWS.WorkDocs.Types.DocumentVersionMetadata

Description

Documentation

data DocumentVersionMetadata

Describes a version of a document.

See: newDocumentVersionMetadata smart constructor.

Constructors

DocumentVersionMetadata'

Fields

modifiedTimestamp :: Maybe POSIX

The timestamp when the document was last uploaded.

status :: Maybe DocumentStatusType

The status of the document.

creatorId :: Maybe Text

The ID of the creator.

contentType :: Maybe Text

The content type of the document.

contentModifiedTimestamp :: Maybe POSIX

The timestamp when the content of the document was modified.

createdTimestamp :: Maybe POSIX

The timestamp when the document was first uploaded.

id :: Maybe Text

The ID of the version.

source :: Maybe (HashMap DocumentSourceType (Sensitive Text))

The source of the document.

name :: Maybe Text

The name of the version.

contentCreatedTimestamp :: Maybe POSIX

The timestamp when the content of the document was originally created.

signature :: Maybe Text

The signature of the document.

thumbnail :: Maybe (HashMap DocumentThumbnailType (Sensitive Text))

The thumbnail of the document.

size :: Maybe Integer

The size of the document, in bytes.

newDocumentVersionMetadata :: DocumentVersionMetadata

Create a value of DocumentVersionMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modifiedTimestamp:DocumentVersionMetadata', documentVersionMetadata_modifiedTimestamp - The timestamp when the document was last uploaded.

$sel:status:DocumentVersionMetadata', documentVersionMetadata_status - The status of the document.

$sel:creatorId:DocumentVersionMetadata', documentVersionMetadata_creatorId - The ID of the creator.

$sel:contentType:DocumentVersionMetadata', documentVersionMetadata_contentType - The content type of the document.

$sel:contentModifiedTimestamp:DocumentVersionMetadata', documentVersionMetadata_contentModifiedTimestamp - The timestamp when the content of the document was modified.

$sel:createdTimestamp:DocumentVersionMetadata', documentVersionMetadata_createdTimestamp - The timestamp when the document was first uploaded.

$sel:id:DocumentVersionMetadata', documentVersionMetadata_id - The ID of the version.

$sel:source:DocumentVersionMetadata', documentVersionMetadata_source - The source of the document.

$sel:name:DocumentVersionMetadata', documentVersionMetadata_name - The name of the version.

$sel:contentCreatedTimestamp:DocumentVersionMetadata', documentVersionMetadata_contentCreatedTimestamp - The timestamp when the content of the document was originally created.

$sel:signature:DocumentVersionMetadata', documentVersionMetadata_signature - The signature of the document.

$sel:thumbnail:DocumentVersionMetadata', documentVersionMetadata_thumbnail - The thumbnail of the document.

$sel:size:DocumentVersionMetadata', documentVersionMetadata_size - The size of the document, in bytes.

documentVersionMetadata_modifiedTimestamp :: Lens' DocumentVersionMetadata (Maybe UTCTime)

The timestamp when the document was last uploaded.

documentVersionMetadata_status :: Lens' DocumentVersionMetadata (Maybe DocumentStatusType)

The status of the document.

documentVersionMetadata_creatorId :: Lens' DocumentVersionMetadata (Maybe Text)

The ID of the creator.

documentVersionMetadata_contentType :: Lens' DocumentVersionMetadata (Maybe Text)

The content type of the document.

documentVersionMetadata_contentModifiedTimestamp :: Lens' DocumentVersionMetadata (Maybe UTCTime)

The timestamp when the content of the document was modified.

documentVersionMetadata_createdTimestamp :: Lens' DocumentVersionMetadata (Maybe UTCTime)

The timestamp when the document was first uploaded.

documentVersionMetadata_id :: Lens' DocumentVersionMetadata (Maybe Text)

The ID of the version.

documentVersionMetadata_source :: Lens' DocumentVersionMetadata (Maybe (HashMap DocumentSourceType Text))

The source of the document.

documentVersionMetadata_name :: Lens' DocumentVersionMetadata (Maybe Text)

The name of the version.

documentVersionMetadata_contentCreatedTimestamp :: Lens' DocumentVersionMetadata (Maybe UTCTime)

The timestamp when the content of the document was originally created.

documentVersionMetadata_signature :: Lens' DocumentVersionMetadata (Maybe Text)

The signature of the document.

documentVersionMetadata_thumbnail :: Lens' DocumentVersionMetadata (Maybe (HashMap DocumentThumbnailType Text))

The thumbnail of the document.

documentVersionMetadata_size :: Lens' DocumentVersionMetadata (Maybe Integer)

The size of the document, in bytes.