Network.AWS.WorkDocs.Types.FolderMetadata

Description

Documentation

data FolderMetadata

Describes a folder.

See: newFolderMetadata smart constructor.

Constructors

FolderMetadata'

Fields

modifiedTimestamp :: Maybe POSIX

The time when the folder was updated.

parentFolderId :: Maybe Text

The ID of the parent folder.

latestVersionSize :: Maybe Integer

The size of the latest version of the folder metadata.

creatorId :: Maybe Text

The ID of the creator.

createdTimestamp :: Maybe POSIX

The time when the folder was created.

labels :: Maybe [Text]

List of labels on the folder.

id :: Maybe Text

The ID of the folder.

name :: Maybe Text

The name of the folder.

signature :: Maybe Text

The unique identifier created from the subfolders and documents of the folder.

size :: Maybe Integer

The size of the folder metadata.

resourceState :: Maybe ResourceStateType

The resource state of the folder.

newFolderMetadata :: FolderMetadata

Create a value of FolderMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modifiedTimestamp:FolderMetadata', folderMetadata_modifiedTimestamp - The time when the folder was updated.

$sel:parentFolderId:FolderMetadata', folderMetadata_parentFolderId - The ID of the parent folder.

$sel:latestVersionSize:FolderMetadata', folderMetadata_latestVersionSize - The size of the latest version of the folder metadata.

$sel:creatorId:FolderMetadata', folderMetadata_creatorId - The ID of the creator.

$sel:createdTimestamp:FolderMetadata', folderMetadata_createdTimestamp - The time when the folder was created.

$sel:labels:FolderMetadata', folderMetadata_labels - List of labels on the folder.

$sel:id:FolderMetadata', folderMetadata_id - The ID of the folder.

$sel:name:FolderMetadata', folderMetadata_name - The name of the folder.

$sel:signature:FolderMetadata', folderMetadata_signature - The unique identifier created from the subfolders and documents of the folder.

$sel:size:FolderMetadata', folderMetadata_size - The size of the folder metadata.

$sel:resourceState:FolderMetadata', folderMetadata_resourceState - The resource state of the folder.

folderMetadata_modifiedTimestamp :: Lens' FolderMetadata (Maybe UTCTime)

The time when the folder was updated.

folderMetadata_parentFolderId :: Lens' FolderMetadata (Maybe Text)

The ID of the parent folder.

folderMetadata_latestVersionSize :: Lens' FolderMetadata (Maybe Integer)

The size of the latest version of the folder metadata.

folderMetadata_creatorId :: Lens' FolderMetadata (Maybe Text)

The ID of the creator.

folderMetadata_createdTimestamp :: Lens' FolderMetadata (Maybe UTCTime)

The time when the folder was created.

folderMetadata_labels :: Lens' FolderMetadata (Maybe [Text])

List of labels on the folder.

folderMetadata_id :: Lens' FolderMetadata (Maybe Text)

The ID of the folder.

folderMetadata_name :: Lens' FolderMetadata (Maybe Text)

The name of the folder.

folderMetadata_signature :: Lens' FolderMetadata (Maybe Text)

The unique identifier created from the subfolders and documents of the folder.

folderMetadata_size :: Lens' FolderMetadata (Maybe Integer)

The size of the folder metadata.

folderMetadata_resourceState :: Lens' FolderMetadata (Maybe ResourceStateType)

The resource state of the folder.