Network.AWS.WorkDocs.Types.Subscription

Description

Documentation

data Subscription

Describes a subscription.

See: newSubscription smart constructor.

Constructors

Subscription'

Fields

subscriptionId :: Maybe Text

The ID of the subscription.

protocol :: Maybe SubscriptionProtocolType

The protocol of the subscription.

endPoint :: Maybe Text

The endpoint of the subscription.

newSubscription :: Subscription

Create a value of Subscription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscriptionId:Subscription', subscription_subscriptionId - The ID of the subscription.

$sel:protocol:Subscription', subscription_protocol - The protocol of the subscription.

$sel:endPoint:Subscription', subscription_endPoint - The endpoint of the subscription.

subscription_subscriptionId :: Lens' Subscription (Maybe Text)

The ID of the subscription.

subscription_protocol :: Lens' Subscription (Maybe SubscriptionProtocolType)

The protocol of the subscription.

subscription_endPoint :: Lens' Subscription (Maybe Text)

The endpoint of the subscription.