Network.AWS.WorkDocs.Types.User

Description

Documentation

data User

Describes a user.

See: newUser smart constructor.

Constructors

User'

Fields

modifiedTimestamp :: Maybe POSIX

The time when the user was modified.

status :: Maybe UserStatusType

The status of the user.

organizationId :: Maybe Text

The ID of the organization.

timeZoneId :: Maybe Text

The time zone ID of the user.

createdTimestamp :: Maybe POSIX

The time when the user was created.

surname :: Maybe Text

The surname of the user.

locale :: Maybe LocaleType

The locale of the user.

rootFolderId :: Maybe Text

The ID of the root folder.

id :: Maybe Text

The ID of the user.

givenName :: Maybe Text

The given name of the user.

recycleBinFolderId :: Maybe Text

The ID of the recycle bin folder.

storage :: Maybe UserStorageMetadata

The storage for the user.

username :: Maybe Text

The login name of the user.

type' :: Maybe UserType

The type of user.

emailAddress :: Maybe Text

The email address of the user.

newUser :: User

Create a value of User with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modifiedTimestamp:User', user_modifiedTimestamp - The time when the user was modified.

$sel:status:User', user_status - The status of the user.

$sel:organizationId:User', user_organizationId - The ID of the organization.

$sel:timeZoneId:User', user_timeZoneId - The time zone ID of the user.

$sel:createdTimestamp:User', user_createdTimestamp - The time when the user was created.

$sel:surname:User', user_surname - The surname of the user.

$sel:locale:User', user_locale - The locale of the user.

$sel:rootFolderId:User', user_rootFolderId - The ID of the root folder.

$sel:id:User', user_id - The ID of the user.

$sel:givenName:User', user_givenName - The given name of the user.

$sel:recycleBinFolderId:User', user_recycleBinFolderId - The ID of the recycle bin folder.

$sel:storage:User', user_storage - The storage for the user.

$sel:username:User', user_username - The login name of the user.

$sel:type':User', user_type - The type of user.

$sel:emailAddress:User', user_emailAddress - The email address of the user.

user_modifiedTimestamp :: Lens' User (Maybe UTCTime)

The time when the user was modified.

user_status :: Lens' User (Maybe UserStatusType)

The status of the user.

user_organizationId :: Lens' User (Maybe Text)

The ID of the organization.

user_timeZoneId :: Lens' User (Maybe Text)

The time zone ID of the user.

user_createdTimestamp :: Lens' User (Maybe UTCTime)

The time when the user was created.

user_surname :: Lens' User (Maybe Text)

The surname of the user.

user_locale :: Lens' User (Maybe LocaleType)

The locale of the user.

user_rootFolderId :: Lens' User (Maybe Text)

The ID of the root folder.

user_id :: Lens' User (Maybe Text)

The ID of the user.

user_givenName :: Lens' User (Maybe Text)

The given name of the user.

user_recycleBinFolderId :: Lens' User (Maybe Text)

The ID of the recycle bin folder.

user_storage :: Lens' User (Maybe UserStorageMetadata)

The storage for the user.

user_username :: Lens' User (Maybe Text)

The login name of the user.

user_type :: Lens' User (Maybe UserType)

The type of user.

user_emailAddress :: Lens' User (Maybe Text)

The email address of the user.