Network.AWS.WorkDocs.UpdateFolder

Description

Updates the specified attributes of the specified folder. The user must have access to both the folder and its parent folder, if applicable.

Creating a Request

data UpdateFolder

See: newUpdateFolder smart constructor.

Constructors

UpdateFolder'

Fields

parentFolderId :: Maybe Text

The ID of the parent folder.

name :: Maybe Text

The name of the folder.

authenticationToken :: Maybe (Sensitive Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

resourceState :: Maybe ResourceStateType

The resource state of the folder. Only ACTIVE and RECYCLED are accepted values from the API.

folderId :: Text

The ID of the folder.

newUpdateFolder

Arguments

:: Text
$sel:folderId:UpdateFolder'

-> UpdateFolder

Create a value of UpdateFolder with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parentFolderId:UpdateFolder', updateFolder_parentFolderId - The ID of the parent folder.

$sel:name:UpdateFolder', updateFolder_name - The name of the folder.

$sel:authenticationToken:UpdateFolder', updateFolder_authenticationToken - Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

$sel:resourceState:UpdateFolder', updateFolder_resourceState - The resource state of the folder. Only ACTIVE and RECYCLED are accepted values from the API.

$sel:folderId:UpdateFolder', updateFolder_folderId - The ID of the folder.

Request Lenses

updateFolder_parentFolderId :: Lens' UpdateFolder (Maybe Text)

The ID of the parent folder.

updateFolder_name :: Lens' UpdateFolder (Maybe Text)

The name of the folder.

updateFolder_authenticationToken :: Lens' UpdateFolder (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

updateFolder_resourceState :: Lens' UpdateFolder (Maybe ResourceStateType)

The resource state of the folder. Only ACTIVE and RECYCLED are accepted values from the API.

updateFolder_folderId :: Lens' UpdateFolder Text

The ID of the folder.

Destructuring the Response

data UpdateFolderResponse

See: newUpdateFolderResponse smart constructor.

Constructors

UpdateFolderResponse'

newUpdateFolderResponse :: UpdateFolderResponse

Create a value of UpdateFolderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.