Network.AWS.WorkDocs.UpdateUser

Description

Updates the specified attributes of the specified user, and grants or revokes administrative privileges to the Amazon WorkDocs site.

Creating a Request

data UpdateUser

See: newUpdateUser smart constructor.

Constructors

UpdateUser'

Fields

storageRule :: Maybe StorageRuleType

The amount of storage for the user.

grantPoweruserPrivileges :: Maybe BooleanEnumType

Boolean value to determine whether the user is granted Poweruser privileges.

timeZoneId :: Maybe Text

The time zone ID of the user.

surname :: Maybe Text

The surname of the user.

locale :: Maybe LocaleType

The locale of the user.

givenName :: Maybe Text

The given name of the user.

authenticationToken :: Maybe (Sensitive Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

type' :: Maybe UserType

The type of the user.

userId :: Text

The ID of the user.

newUpdateUser

Arguments

:: Text
$sel:userId:UpdateUser'

-> UpdateUser

Create a value of UpdateUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:storageRule:UpdateUser', updateUser_storageRule - The amount of storage for the user.

$sel:grantPoweruserPrivileges:UpdateUser', updateUser_grantPoweruserPrivileges - Boolean value to determine whether the user is granted Poweruser privileges.

$sel:timeZoneId:UpdateUser', updateUser_timeZoneId - The time zone ID of the user.

$sel:surname:UpdateUser', updateUser_surname - The surname of the user.

$sel:locale:UpdateUser', updateUser_locale - The locale of the user.

$sel:givenName:UpdateUser', updateUser_givenName - The given name of the user.

$sel:authenticationToken:UpdateUser', updateUser_authenticationToken - Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

$sel:type':UpdateUser', updateUser_type - The type of the user.

$sel:userId:UpdateUser', updateUser_userId - The ID of the user.

Request Lenses

updateUser_storageRule :: Lens' UpdateUser (Maybe StorageRuleType)

The amount of storage for the user.

updateUser_grantPoweruserPrivileges :: Lens' UpdateUser (Maybe BooleanEnumType)

Boolean value to determine whether the user is granted Poweruser privileges.

updateUser_timeZoneId :: Lens' UpdateUser (Maybe Text)

The time zone ID of the user.

updateUser_surname :: Lens' UpdateUser (Maybe Text)

The surname of the user.

updateUser_locale :: Lens' UpdateUser (Maybe LocaleType)

The locale of the user.

updateUser_givenName :: Lens' UpdateUser (Maybe Text)

The given name of the user.

updateUser_authenticationToken :: Lens' UpdateUser (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

updateUser_type :: Lens' UpdateUser (Maybe UserType)

The type of the user.

updateUser_userId :: Lens' UpdateUser Text

The ID of the user.

Destructuring the Response

data UpdateUserResponse

See: newUpdateUserResponse smart constructor.

Constructors

UpdateUserResponse'

Fields

user :: Maybe User

The user information.

httpStatus :: Int

The response's http status code.

newUpdateUserResponse

Arguments

:: Int
$sel:httpStatus:UpdateUserResponse'

-> UpdateUserResponse

Create a value of UpdateUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:user:UpdateUserResponse', updateUserResponse_user - The user information.

$sel:httpStatus:UpdateUserResponse', updateUserResponse_httpStatus - The response's http status code.

Response Lenses

updateUserResponse_user :: Lens' UpdateUserResponse (Maybe User)

The user information.

updateUserResponse_httpStatus :: Lens' UpdateUserResponse Int

The response's http status code.