Network.AWS.WorkLink.AssociateWebsiteAuthorizationProvider

Description

Associates a website authorization provider with a specified fleet. This is used to authorize users against associated websites in the company network.

Creating a Request

data AssociateWebsiteAuthorizationProvider

See: newAssociateWebsiteAuthorizationProvider smart constructor.

Constructors

AssociateWebsiteAuthorizationProvider'

Fields

domainName :: Maybe Text

The domain name of the authorization provider. This applies only to SAML-based authorization providers.

fleetArn :: Text

The ARN of the fleet.

authorizationProviderType :: AuthorizationProviderType

The authorization provider type.

newAssociateWebsiteAuthorizationProvider

Arguments

:: Text
$sel:fleetArn:AssociateWebsiteAuthorizationProvider'

-> AuthorizationProviderType
$sel:authorizationProviderType:AssociateWebsiteAuthorizationProvider'

-> AssociateWebsiteAuthorizationProvider

Create a value of AssociateWebsiteAuthorizationProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:AssociateWebsiteAuthorizationProvider', associateWebsiteAuthorizationProvider_domainName - The domain name of the authorization provider. This applies only to SAML-based authorization providers.

$sel:fleetArn:AssociateWebsiteAuthorizationProvider', associateWebsiteAuthorizationProvider_fleetArn - The ARN of the fleet.

$sel:authorizationProviderType:AssociateWebsiteAuthorizationProvider', associateWebsiteAuthorizationProvider_authorizationProviderType - The authorization provider type.

Request Lenses

associateWebsiteAuthorizationProvider_domainName :: Lens' AssociateWebsiteAuthorizationProvider (Maybe Text)

The domain name of the authorization provider. This applies only to SAML-based authorization providers.

associateWebsiteAuthorizationProvider_fleetArn :: Lens' AssociateWebsiteAuthorizationProvider Text

The ARN of the fleet.

associateWebsiteAuthorizationProvider_authorizationProviderType :: Lens' AssociateWebsiteAuthorizationProvider AuthorizationProviderType

The authorization provider type.

Destructuring the Response

data AssociateWebsiteAuthorizationProviderResponse

See: newAssociateWebsiteAuthorizationProviderResponse smart constructor.

Constructors

AssociateWebsiteAuthorizationProviderResponse'

Fields

authorizationProviderId :: Maybe Text

A unique identifier for the authorization provider.

httpStatus :: Int

The response's http status code.

newAssociateWebsiteAuthorizationProviderResponse

Arguments

:: Int
$sel:httpStatus:AssociateWebsiteAuthorizationProviderResponse'

-> AssociateWebsiteAuthorizationProviderResponse

Create a value of AssociateWebsiteAuthorizationProviderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorizationProviderId:AssociateWebsiteAuthorizationProviderResponse', associateWebsiteAuthorizationProviderResponse_authorizationProviderId - A unique identifier for the authorization provider.

$sel:httpStatus:AssociateWebsiteAuthorizationProviderResponse', associateWebsiteAuthorizationProviderResponse_httpStatus - The response's http status code.

Response Lenses

associateWebsiteAuthorizationProviderResponse_authorizationProviderId :: Lens' AssociateWebsiteAuthorizationProviderResponse (Maybe Text)

A unique identifier for the authorization provider.

associateWebsiteAuthorizationProviderResponse_httpStatus :: Lens' AssociateWebsiteAuthorizationProviderResponse Int

The response's http status code.