Network.AWS.WorkLink.AssociateWebsiteCertificateAuthority

Description

Imports the root certificate of a certificate authority (CA) used to obtain TLS certificates used by associated websites within the company network.

Creating a Request

data AssociateWebsiteCertificateAuthority

See: newAssociateWebsiteCertificateAuthority smart constructor.

Constructors

AssociateWebsiteCertificateAuthority'

Fields

displayName :: Maybe Text

The certificate name to display.

fleetArn :: Text

The ARN of the fleet.

certificate :: Text

The root certificate of the CA.

newAssociateWebsiteCertificateAuthority

Arguments

:: Text
$sel:fleetArn:AssociateWebsiteCertificateAuthority'

-> Text
$sel:certificate:AssociateWebsiteCertificateAuthority'

-> AssociateWebsiteCertificateAuthority

Create a value of AssociateWebsiteCertificateAuthority with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:displayName:AssociateWebsiteCertificateAuthority', associateWebsiteCertificateAuthority_displayName - The certificate name to display.

$sel:fleetArn:AssociateWebsiteCertificateAuthority', associateWebsiteCertificateAuthority_fleetArn - The ARN of the fleet.

$sel:certificate:AssociateWebsiteCertificateAuthority', associateWebsiteCertificateAuthority_certificate - The root certificate of the CA.

Request Lenses

associateWebsiteCertificateAuthority_displayName :: Lens' AssociateWebsiteCertificateAuthority (Maybe Text)

The certificate name to display.

associateWebsiteCertificateAuthority_fleetArn :: Lens' AssociateWebsiteCertificateAuthority Text

The ARN of the fleet.

associateWebsiteCertificateAuthority_certificate :: Lens' AssociateWebsiteCertificateAuthority Text

The root certificate of the CA.

Destructuring the Response

data AssociateWebsiteCertificateAuthorityResponse

See: newAssociateWebsiteCertificateAuthorityResponse smart constructor.

Constructors

AssociateWebsiteCertificateAuthorityResponse'

Fields

websiteCaId :: Maybe Text

A unique identifier for the CA.

httpStatus :: Int

The response's http status code.

newAssociateWebsiteCertificateAuthorityResponse

Arguments

:: Int
$sel:httpStatus:AssociateWebsiteCertificateAuthorityResponse'

-> AssociateWebsiteCertificateAuthorityResponse

Create a value of AssociateWebsiteCertificateAuthorityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:websiteCaId:AssociateWebsiteCertificateAuthorityResponse', associateWebsiteCertificateAuthorityResponse_websiteCaId - A unique identifier for the CA.

$sel:httpStatus:AssociateWebsiteCertificateAuthorityResponse', associateWebsiteCertificateAuthorityResponse_httpStatus - The response's http status code.

Response Lenses

associateWebsiteCertificateAuthorityResponse_websiteCaId :: Lens' AssociateWebsiteCertificateAuthorityResponse (Maybe Text)

A unique identifier for the CA.

associateWebsiteCertificateAuthorityResponse_httpStatus :: Lens' AssociateWebsiteCertificateAuthorityResponse Int

The response's http status code.