Network.AWS.WorkLink.DescribeDevicePolicyConfiguration

Description

Describes the device policy configuration for the specified fleet.

Creating a Request

data DescribeDevicePolicyConfiguration

See: newDescribeDevicePolicyConfiguration smart constructor.

Constructors

DescribeDevicePolicyConfiguration'

Fields

fleetArn :: Text

The ARN of the fleet.

newDescribeDevicePolicyConfiguration

Arguments

:: Text
$sel:fleetArn:DescribeDevicePolicyConfiguration'

-> DescribeDevicePolicyConfiguration

Create a value of DescribeDevicePolicyConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetArn:DescribeDevicePolicyConfiguration', describeDevicePolicyConfiguration_fleetArn - The ARN of the fleet.

Request Lenses

describeDevicePolicyConfiguration_fleetArn :: Lens' DescribeDevicePolicyConfiguration Text

The ARN of the fleet.

Destructuring the Response

data DescribeDevicePolicyConfigurationResponse

See: newDescribeDevicePolicyConfigurationResponse smart constructor.

Constructors

DescribeDevicePolicyConfigurationResponse'

Fields

deviceCaCertificate :: Maybe Text

The certificate chain, including intermediate certificates and the root certificate authority certificate used to issue device certificates.

httpStatus :: Int

The response's http status code.

newDescribeDevicePolicyConfigurationResponse

Arguments

:: Int
$sel:httpStatus:DescribeDevicePolicyConfigurationResponse'

-> DescribeDevicePolicyConfigurationResponse

Create a value of DescribeDevicePolicyConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceCaCertificate:DescribeDevicePolicyConfigurationResponse', describeDevicePolicyConfigurationResponse_deviceCaCertificate - The certificate chain, including intermediate certificates and the root certificate authority certificate used to issue device certificates.

$sel:httpStatus:DescribeDevicePolicyConfigurationResponse', describeDevicePolicyConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

describeDevicePolicyConfigurationResponse_deviceCaCertificate :: Lens' DescribeDevicePolicyConfigurationResponse (Maybe Text)

The certificate chain, including intermediate certificates and the root certificate authority certificate used to issue device certificates.

describeDevicePolicyConfigurationResponse_httpStatus :: Lens' DescribeDevicePolicyConfigurationResponse Int

The response's http status code.