Network.AWS.WorkLink.DescribeIdentityProviderConfiguration

Description

Describes the identity provider configuration of the specified fleet.

Creating a Request

data DescribeIdentityProviderConfiguration

See: newDescribeIdentityProviderConfiguration smart constructor.

Constructors

DescribeIdentityProviderConfiguration'

Fields

fleetArn :: Text

The ARN of the fleet.

newDescribeIdentityProviderConfiguration

Arguments

:: Text
$sel:fleetArn:DescribeIdentityProviderConfiguration'

-> DescribeIdentityProviderConfiguration

Create a value of DescribeIdentityProviderConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetArn:DescribeIdentityProviderConfiguration', describeIdentityProviderConfiguration_fleetArn - The ARN of the fleet.

Request Lenses

describeIdentityProviderConfiguration_fleetArn :: Lens' DescribeIdentityProviderConfiguration Text

The ARN of the fleet.

Destructuring the Response

data DescribeIdentityProviderConfigurationResponse

See: newDescribeIdentityProviderConfigurationResponse smart constructor.

Constructors

DescribeIdentityProviderConfigurationResponse'

Fields

identityProviderType :: Maybe IdentityProviderType

The type of identity provider.

serviceProviderSamlMetadata :: Maybe Text

The SAML metadata document uploaded to the user’s identity provider.

identityProviderSamlMetadata :: Maybe Text

The SAML metadata document provided by the user’s identity provider.

httpStatus :: Int

The response's http status code.

newDescribeIdentityProviderConfigurationResponse

Arguments

:: Int
$sel:httpStatus:DescribeIdentityProviderConfigurationResponse'

-> DescribeIdentityProviderConfigurationResponse

Create a value of DescribeIdentityProviderConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityProviderType:DescribeIdentityProviderConfigurationResponse', describeIdentityProviderConfigurationResponse_identityProviderType - The type of identity provider.

$sel:serviceProviderSamlMetadata:DescribeIdentityProviderConfigurationResponse', describeIdentityProviderConfigurationResponse_serviceProviderSamlMetadata - The SAML metadata document uploaded to the user’s identity provider.

$sel:identityProviderSamlMetadata:DescribeIdentityProviderConfigurationResponse', describeIdentityProviderConfigurationResponse_identityProviderSamlMetadata - The SAML metadata document provided by the user’s identity provider.

$sel:httpStatus:DescribeIdentityProviderConfigurationResponse', describeIdentityProviderConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

describeIdentityProviderConfigurationResponse_identityProviderType :: Lens' DescribeIdentityProviderConfigurationResponse (Maybe IdentityProviderType)

The type of identity provider.

describeIdentityProviderConfigurationResponse_serviceProviderSamlMetadata :: Lens' DescribeIdentityProviderConfigurationResponse (Maybe Text)

The SAML metadata document uploaded to the user’s identity provider.

describeIdentityProviderConfigurationResponse_identityProviderSamlMetadata :: Lens' DescribeIdentityProviderConfigurationResponse (Maybe Text)

The SAML metadata document provided by the user’s identity provider.

describeIdentityProviderConfigurationResponse_httpStatus :: Lens' DescribeIdentityProviderConfigurationResponse Int

The response's http status code.