Network.AWS.WorkLink.DescribeWebsiteCertificateAuthority

Description

Provides information about the certificate authority.

Creating a Request

data DescribeWebsiteCertificateAuthority

See: newDescribeWebsiteCertificateAuthority smart constructor.

Constructors

DescribeWebsiteCertificateAuthority'

Fields

fleetArn :: Text

The ARN of the fleet.

websiteCaId :: Text

A unique identifier for the certificate authority.

newDescribeWebsiteCertificateAuthority

Arguments

:: Text
$sel:fleetArn:DescribeWebsiteCertificateAuthority'

-> Text
$sel:websiteCaId:DescribeWebsiteCertificateAuthority'

-> DescribeWebsiteCertificateAuthority

Create a value of DescribeWebsiteCertificateAuthority with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetArn:DescribeWebsiteCertificateAuthority', describeWebsiteCertificateAuthority_fleetArn - The ARN of the fleet.

$sel:websiteCaId:DescribeWebsiteCertificateAuthority', describeWebsiteCertificateAuthority_websiteCaId - A unique identifier for the certificate authority.

Request Lenses

describeWebsiteCertificateAuthority_fleetArn :: Lens' DescribeWebsiteCertificateAuthority Text

The ARN of the fleet.

describeWebsiteCertificateAuthority_websiteCaId :: Lens' DescribeWebsiteCertificateAuthority Text

A unique identifier for the certificate authority.

Destructuring the Response

data DescribeWebsiteCertificateAuthorityResponse

See: newDescribeWebsiteCertificateAuthorityResponse smart constructor.

Constructors

DescribeWebsiteCertificateAuthorityResponse'

Fields

createdTime :: Maybe POSIX

The time that the certificate authority was added.

certificate :: Maybe Text

The root certificate of the certificate authority.

displayName :: Maybe Text

The certificate name to display.

httpStatus :: Int

The response's http status code.

newDescribeWebsiteCertificateAuthorityResponse

Arguments

:: Int
$sel:httpStatus:DescribeWebsiteCertificateAuthorityResponse'

-> DescribeWebsiteCertificateAuthorityResponse

Create a value of DescribeWebsiteCertificateAuthorityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdTime:DescribeWebsiteCertificateAuthorityResponse', describeWebsiteCertificateAuthorityResponse_createdTime - The time that the certificate authority was added.

$sel:certificate:DescribeWebsiteCertificateAuthorityResponse', describeWebsiteCertificateAuthorityResponse_certificate - The root certificate of the certificate authority.

$sel:displayName:DescribeWebsiteCertificateAuthorityResponse', describeWebsiteCertificateAuthorityResponse_displayName - The certificate name to display.

$sel:httpStatus:DescribeWebsiteCertificateAuthorityResponse', describeWebsiteCertificateAuthorityResponse_httpStatus - The response's http status code.

Response Lenses

describeWebsiteCertificateAuthorityResponse_createdTime :: Lens' DescribeWebsiteCertificateAuthorityResponse (Maybe UTCTime)

The time that the certificate authority was added.

describeWebsiteCertificateAuthorityResponse_certificate :: Lens' DescribeWebsiteCertificateAuthorityResponse (Maybe Text)

The root certificate of the certificate authority.

describeWebsiteCertificateAuthorityResponse_displayName :: Lens' DescribeWebsiteCertificateAuthorityResponse (Maybe Text)

The certificate name to display.

describeWebsiteCertificateAuthorityResponse_httpStatus :: Lens' DescribeWebsiteCertificateAuthorityResponse Int

The response's http status code.