Network.AWS.WorkLink.Lens

Description

Operations

UpdateDomainMetadata

updateDomainMetadata_displayName :: Lens' UpdateDomainMetadata (Maybe Text)

The name to display.

updateDomainMetadata_fleetArn :: Lens' UpdateDomainMetadata Text

The ARN of the fleet.

updateDomainMetadata_domainName :: Lens' UpdateDomainMetadata Text

The name of the domain.

updateDomainMetadataResponse_httpStatus :: Lens' UpdateDomainMetadataResponse Int

The response's http status code.

SignOutUser

signOutUser_fleetArn :: Lens' SignOutUser Text

The ARN of the fleet.

signOutUser_username :: Lens' SignOutUser Text

The name of the user.

signOutUserResponse_httpStatus :: Lens' SignOutUserResponse Int

The response's http status code.

RestoreDomainAccess

restoreDomainAccess_fleetArn :: Lens' RestoreDomainAccess Text

The ARN of the fleet.

restoreDomainAccess_domainName :: Lens' RestoreDomainAccess Text

The name of the domain.

restoreDomainAccessResponse_httpStatus :: Lens' RestoreDomainAccessResponse Int

The response's http status code.

AssociateDomain

associateDomain_displayName :: Lens' AssociateDomain (Maybe Text)

The name to display.

associateDomain_fleetArn :: Lens' AssociateDomain Text

The Amazon Resource Name (ARN) of the fleet.

associateDomain_domainName :: Lens' AssociateDomain Text

The fully qualified domain name (FQDN).

associateDomain_acmCertificateArn :: Lens' AssociateDomain Text

The ARN of an issued ACM certificate that is valid for the domain being associated.

associateDomainResponse_httpStatus :: Lens' AssociateDomainResponse Int

The response's http status code.

DeleteFleet

deleteFleet_fleetArn :: Lens' DeleteFleet Text

The ARN of the fleet.

deleteFleetResponse_httpStatus :: Lens' DeleteFleetResponse Int

The response's http status code.

DisassociateDomain

disassociateDomain_fleetArn :: Lens' DisassociateDomain Text

The ARN of the fleet.

disassociateDomain_domainName :: Lens' DisassociateDomain Text

The name of the domain.

disassociateDomainResponse_httpStatus :: Lens' DisassociateDomainResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the fleet.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The tags attached to the resource. A tag is a key-value pair.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

AssociateWebsiteAuthorizationProvider

associateWebsiteAuthorizationProvider_domainName :: Lens' AssociateWebsiteAuthorizationProvider (Maybe Text)

The domain name of the authorization provider. This applies only to SAML-based authorization providers.

associateWebsiteAuthorizationProvider_fleetArn :: Lens' AssociateWebsiteAuthorizationProvider Text

The ARN of the fleet.

associateWebsiteAuthorizationProvider_authorizationProviderType :: Lens' AssociateWebsiteAuthorizationProvider AuthorizationProviderType

The authorization provider type.

associateWebsiteAuthorizationProviderResponse_authorizationProviderId :: Lens' AssociateWebsiteAuthorizationProviderResponse (Maybe Text)

A unique identifier for the authorization provider.

associateWebsiteAuthorizationProviderResponse_httpStatus :: Lens' AssociateWebsiteAuthorizationProviderResponse Int

The response's http status code.

DescribeDevicePolicyConfiguration

describeDevicePolicyConfiguration_fleetArn :: Lens' DescribeDevicePolicyConfiguration Text

The ARN of the fleet.

describeDevicePolicyConfigurationResponse_deviceCaCertificate :: Lens' DescribeDevicePolicyConfigurationResponse (Maybe Text)

The certificate chain, including intermediate certificates and the root certificate authority certificate used to issue device certificates.

describeDevicePolicyConfigurationResponse_httpStatus :: Lens' DescribeDevicePolicyConfigurationResponse Int

The response's http status code.

ListWebsiteAuthorizationProviders

listWebsiteAuthorizationProviders_nextToken :: Lens' ListWebsiteAuthorizationProviders (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

listWebsiteAuthorizationProviders_maxResults :: Lens' ListWebsiteAuthorizationProviders (Maybe Natural)

The maximum number of results to be included in the next page.

listWebsiteAuthorizationProviders_fleetArn :: Lens' ListWebsiteAuthorizationProviders Text

The ARN of the fleet.

listWebsiteAuthorizationProvidersResponse_websiteAuthorizationProviders :: Lens' ListWebsiteAuthorizationProvidersResponse (Maybe [WebsiteAuthorizationProviderSummary])

The website authorization providers.

listWebsiteAuthorizationProvidersResponse_nextToken :: Lens' ListWebsiteAuthorizationProvidersResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

listWebsiteAuthorizationProvidersResponse_httpStatus :: Lens' ListWebsiteAuthorizationProvidersResponse Int

The response's http status code.

DisassociateWebsiteAuthorizationProvider

disassociateWebsiteAuthorizationProvider_fleetArn :: Lens' DisassociateWebsiteAuthorizationProvider Text

The ARN of the fleet.

disassociateWebsiteAuthorizationProvider_authorizationProviderId :: Lens' DisassociateWebsiteAuthorizationProvider Text

A unique identifier for the authorization provider.

disassociateWebsiteAuthorizationProviderResponse_httpStatus :: Lens' DisassociateWebsiteAuthorizationProviderResponse Int

The response's http status code.

ListWebsiteCertificateAuthorities

listWebsiteCertificateAuthorities_nextToken :: Lens' ListWebsiteCertificateAuthorities (Maybe Text)

The pagination token used to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

listWebsiteCertificateAuthorities_maxResults :: Lens' ListWebsiteCertificateAuthorities (Maybe Natural)

The maximum number of results to be included in the next page.

listWebsiteCertificateAuthorities_fleetArn :: Lens' ListWebsiteCertificateAuthorities Text

The ARN of the fleet.

listWebsiteCertificateAuthoritiesResponse_websiteCertificateAuthorities :: Lens' ListWebsiteCertificateAuthoritiesResponse (Maybe [WebsiteCaSummary])

Information about the certificates.

listWebsiteCertificateAuthoritiesResponse_nextToken :: Lens' ListWebsiteCertificateAuthoritiesResponse (Maybe Text)

The pagination token used to retrieve the next page of results for this operation. If there are no more pages, this value is null.

listWebsiteCertificateAuthoritiesResponse_httpStatus :: Lens' ListWebsiteCertificateAuthoritiesResponse Int

The response's http status code.

UpdateIdentityProviderConfiguration

updateIdentityProviderConfiguration_identityProviderSamlMetadata :: Lens' UpdateIdentityProviderConfiguration (Maybe Text)

The SAML metadata document provided by the customer’s identity provider. The existing IdentityProviderSamlMetadata is unset if null is passed.

updateIdentityProviderConfiguration_fleetArn :: Lens' UpdateIdentityProviderConfiguration Text

The ARN of the fleet.

updateIdentityProviderConfiguration_identityProviderType :: Lens' UpdateIdentityProviderConfiguration IdentityProviderType

The type of identity provider.

updateIdentityProviderConfigurationResponse_httpStatus :: Lens' UpdateIdentityProviderConfigurationResponse Int

The response's http status code.

DescribeFleetMetadata

describeFleetMetadata_fleetArn :: Lens' DescribeFleetMetadata Text

The Amazon Resource Name (ARN) of the fleet.

describeFleetMetadataResponse_lastUpdatedTime :: Lens' DescribeFleetMetadataResponse (Maybe UTCTime)

The time that the fleet was last updated.

describeFleetMetadataResponse_fleetStatus :: Lens' DescribeFleetMetadataResponse (Maybe FleetStatus)

The current state of the fleet.

describeFleetMetadataResponse_companyCode :: Lens' DescribeFleetMetadataResponse (Maybe Text)

The identifier used by users to sign in to the Amazon WorkLink app.

describeFleetMetadataResponse_createdTime :: Lens' DescribeFleetMetadataResponse (Maybe UTCTime)

The time that the fleet was created.

describeFleetMetadataResponse_optimizeForEndUserLocation :: Lens' DescribeFleetMetadataResponse (Maybe Bool)

The option to optimize for better performance by routing traffic through the closest AWS Region to users, which may be outside of your home Region.

describeFleetMetadataResponse_displayName :: Lens' DescribeFleetMetadataResponse (Maybe Text)

The name to display.

describeFleetMetadataResponse_fleetName :: Lens' DescribeFleetMetadataResponse (Maybe Text)

The name of the fleet.

describeFleetMetadataResponse_tags :: Lens' DescribeFleetMetadataResponse (Maybe (HashMap Text Text))

The tags attached to the resource. A tag is a key-value pair.

describeFleetMetadataResponse_httpStatus :: Lens' DescribeFleetMetadataResponse Int

The response's http status code.

UpdateDevicePolicyConfiguration

updateDevicePolicyConfiguration_deviceCaCertificate :: Lens' UpdateDevicePolicyConfiguration (Maybe Text)

The certificate chain, including intermediate certificates and the root certificate authority certificate used to issue device certificates.

updateDevicePolicyConfiguration_fleetArn :: Lens' UpdateDevicePolicyConfiguration Text

The ARN of the fleet.

updateDevicePolicyConfigurationResponse_httpStatus :: Lens' UpdateDevicePolicyConfigurationResponse Int

The response's http status code.

DescribeCompanyNetworkConfiguration

describeCompanyNetworkConfiguration_fleetArn :: Lens' DescribeCompanyNetworkConfiguration Text

The ARN of the fleet.

describeCompanyNetworkConfigurationResponse_securityGroupIds :: Lens' DescribeCompanyNetworkConfigurationResponse (Maybe [Text])

The security groups associated with access to the provided subnets.

describeCompanyNetworkConfigurationResponse_subnetIds :: Lens' DescribeCompanyNetworkConfigurationResponse (Maybe [Text])

The subnets used for X-ENI connections from Amazon WorkLink rendering containers.

describeCompanyNetworkConfigurationResponse_vpcId :: Lens' DescribeCompanyNetworkConfigurationResponse (Maybe Text)

The VPC with connectivity to associated websites.

describeCompanyNetworkConfigurationResponse_httpStatus :: Lens' DescribeCompanyNetworkConfigurationResponse Int

The response's http status code.

DescribeAuditStreamConfiguration

describeAuditStreamConfiguration_fleetArn :: Lens' DescribeAuditStreamConfiguration Text

The ARN of the fleet.

describeAuditStreamConfigurationResponse_auditStreamArn :: Lens' DescribeAuditStreamConfigurationResponse (Maybe Text)

The ARN of the Amazon Kinesis data stream that will receive the audit events.

describeAuditStreamConfigurationResponse_httpStatus :: Lens' DescribeAuditStreamConfigurationResponse Int

The response's http status code.

RevokeDomainAccess

revokeDomainAccess_fleetArn :: Lens' RevokeDomainAccess Text

The ARN of the fleet.

revokeDomainAccess_domainName :: Lens' RevokeDomainAccess Text

The name of the domain.

revokeDomainAccessResponse_httpStatus :: Lens' RevokeDomainAccessResponse Int

The response's http status code.

ListFleets

listFleets_nextToken :: Lens' ListFleets (Maybe Text)

The pagination token used to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

listFleets_maxResults :: Lens' ListFleets (Maybe Natural)

The maximum number of results to be included in the next page.

listFleetsResponse_fleetSummaryList :: Lens' ListFleetsResponse (Maybe [FleetSummary])

The summary list of the fleets.

listFleetsResponse_nextToken :: Lens' ListFleetsResponse (Maybe Text)

The pagination token used to retrieve the next page of results for this operation. If there are no more pages, this value is null.

listFleetsResponse_httpStatus :: Lens' ListFleetsResponse Int

The response's http status code.

UpdateFleetMetadata

updateFleetMetadata_optimizeForEndUserLocation :: Lens' UpdateFleetMetadata (Maybe Bool)

The option to optimize for better performance by routing traffic through the closest AWS Region to users, which may be outside of your home Region.

updateFleetMetadata_displayName :: Lens' UpdateFleetMetadata (Maybe Text)

The fleet name to display. The existing DisplayName is unset if null is passed.

updateFleetMetadata_fleetArn :: Lens' UpdateFleetMetadata Text

The ARN of the fleet.

updateFleetMetadataResponse_httpStatus :: Lens' UpdateFleetMetadataResponse Int

The response's http status code.

DescribeIdentityProviderConfiguration

describeIdentityProviderConfiguration_fleetArn :: Lens' DescribeIdentityProviderConfiguration Text

The ARN of the fleet.

describeIdentityProviderConfigurationResponse_identityProviderType :: Lens' DescribeIdentityProviderConfigurationResponse (Maybe IdentityProviderType)

The type of identity provider.

describeIdentityProviderConfigurationResponse_serviceProviderSamlMetadata :: Lens' DescribeIdentityProviderConfigurationResponse (Maybe Text)

The SAML metadata document uploaded to the user’s identity provider.

describeIdentityProviderConfigurationResponse_identityProviderSamlMetadata :: Lens' DescribeIdentityProviderConfigurationResponse (Maybe Text)

The SAML metadata document provided by the user’s identity provider.

describeIdentityProviderConfigurationResponse_httpStatus :: Lens' DescribeIdentityProviderConfigurationResponse Int

The response's http status code.

CreateFleet

createFleet_optimizeForEndUserLocation :: Lens' CreateFleet (Maybe Bool)

The option to optimize for better performance by routing traffic through the closest AWS Region to users, which may be outside of your home Region.

createFleet_displayName :: Lens' CreateFleet (Maybe Text)

The fleet name to display.

createFleet_tags :: Lens' CreateFleet (Maybe (HashMap Text Text))

The tags to add to the resource. A tag is a key-value pair.

createFleet_fleetName :: Lens' CreateFleet Text

A unique name for the fleet.

createFleetResponse_fleetArn :: Lens' CreateFleetResponse (Maybe Text)

The Amazon Resource Name (ARN) of the fleet.

createFleetResponse_httpStatus :: Lens' CreateFleetResponse Int

The response's http status code.

DescribeWebsiteCertificateAuthority

describeWebsiteCertificateAuthority_fleetArn :: Lens' DescribeWebsiteCertificateAuthority Text

The ARN of the fleet.

describeWebsiteCertificateAuthority_websiteCaId :: Lens' DescribeWebsiteCertificateAuthority Text

A unique identifier for the certificate authority.

describeWebsiteCertificateAuthorityResponse_createdTime :: Lens' DescribeWebsiteCertificateAuthorityResponse (Maybe UTCTime)

The time that the certificate authority was added.

describeWebsiteCertificateAuthorityResponse_certificate :: Lens' DescribeWebsiteCertificateAuthorityResponse (Maybe Text)

The root certificate of the certificate authority.

describeWebsiteCertificateAuthorityResponse_displayName :: Lens' DescribeWebsiteCertificateAuthorityResponse (Maybe Text)

The certificate name to display.

describeWebsiteCertificateAuthorityResponse_httpStatus :: Lens' DescribeWebsiteCertificateAuthorityResponse Int

The response's http status code.

DescribeDomain

describeDomain_fleetArn :: Lens' DescribeDomain Text

The ARN of the fleet.

describeDomain_domainName :: Lens' DescribeDomain Text

The name of the domain.

describeDomainResponse_domainStatus :: Lens' DescribeDomainResponse (Maybe DomainStatus)

The current state for the domain.

describeDomainResponse_acmCertificateArn :: Lens' DescribeDomainResponse (Maybe Text)

The ARN of an issued ACM certificate that is valid for the domain being associated.

describeDomainResponse_createdTime :: Lens' DescribeDomainResponse (Maybe UTCTime)

The time that the domain was added.

describeDomainResponse_domainName :: Lens' DescribeDomainResponse (Maybe Text)

The name of the domain.

describeDomainResponse_displayName :: Lens' DescribeDomainResponse (Maybe Text)

The name to display.

describeDomainResponse_httpStatus :: Lens' DescribeDomainResponse Int

The response's http status code.

DisassociateWebsiteCertificateAuthority

disassociateWebsiteCertificateAuthority_fleetArn :: Lens' DisassociateWebsiteCertificateAuthority Text

The ARN of the fleet.

disassociateWebsiteCertificateAuthority_websiteCaId :: Lens' DisassociateWebsiteCertificateAuthority Text

A unique identifier for the CA.

disassociateWebsiteCertificateAuthorityResponse_httpStatus :: Lens' DisassociateWebsiteCertificateAuthorityResponse Int

The response's http status code.

DescribeDevice

describeDevice_fleetArn :: Lens' DescribeDevice Text

The ARN of the fleet.

describeDevice_deviceId :: Lens' DescribeDevice Text

A unique identifier for a registered user's device.

describeDeviceResponse_status :: Lens' DescribeDeviceResponse (Maybe DeviceStatus)

The current state of the device.

describeDeviceResponse_manufacturer :: Lens' DescribeDeviceResponse (Maybe Text)

The manufacturer of the device.

describeDeviceResponse_lastAccessedTime :: Lens' DescribeDeviceResponse (Maybe UTCTime)

The date that the device last accessed Amazon WorkLink.

describeDeviceResponse_operatingSystem :: Lens' DescribeDeviceResponse (Maybe Text)

The operating system of the device.

describeDeviceResponse_username :: Lens' DescribeDeviceResponse (Maybe Text)

The user name associated with the device.

describeDeviceResponse_model :: Lens' DescribeDeviceResponse (Maybe Text)

The model of the device.

describeDeviceResponse_operatingSystemVersion :: Lens' DescribeDeviceResponse (Maybe Text)

The operating system version of the device.

describeDeviceResponse_firstAccessedTime :: Lens' DescribeDeviceResponse (Maybe UTCTime)

The date that the device first signed in to Amazon WorkLink.

describeDeviceResponse_patchLevel :: Lens' DescribeDeviceResponse (Maybe Text)

The operating system patch level of the device.

describeDeviceResponse_httpStatus :: Lens' DescribeDeviceResponse Int

The response's http status code.

UpdateAuditStreamConfiguration

updateAuditStreamConfiguration_auditStreamArn :: Lens' UpdateAuditStreamConfiguration (Maybe Text)

The ARN of the Amazon Kinesis data stream that receives the audit events.

updateAuditStreamConfiguration_fleetArn :: Lens' UpdateAuditStreamConfiguration Text

The ARN of the fleet.

updateAuditStreamConfigurationResponse_httpStatus :: Lens' UpdateAuditStreamConfigurationResponse Int

The response's http status code.

UpdateCompanyNetworkConfiguration

updateCompanyNetworkConfiguration_fleetArn :: Lens' UpdateCompanyNetworkConfiguration Text

The ARN of the fleet.

updateCompanyNetworkConfiguration_vpcId :: Lens' UpdateCompanyNetworkConfiguration Text

The VPC with connectivity to associated websites.

updateCompanyNetworkConfiguration_subnetIds :: Lens' UpdateCompanyNetworkConfiguration [Text]

The subnets used for X-ENI connections from Amazon WorkLink rendering containers.

updateCompanyNetworkConfiguration_securityGroupIds :: Lens' UpdateCompanyNetworkConfiguration [Text]

The security groups associated with access to the provided subnets.

updateCompanyNetworkConfigurationResponse_httpStatus :: Lens' UpdateCompanyNetworkConfigurationResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the fleet.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The tags to add to the resource. A tag is a key-value pair.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the fleet.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

The list of tag keys to remove from the resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

AssociateWebsiteCertificateAuthority

associateWebsiteCertificateAuthority_displayName :: Lens' AssociateWebsiteCertificateAuthority (Maybe Text)

The certificate name to display.

associateWebsiteCertificateAuthority_fleetArn :: Lens' AssociateWebsiteCertificateAuthority Text

The ARN of the fleet.

associateWebsiteCertificateAuthority_certificate :: Lens' AssociateWebsiteCertificateAuthority Text

The root certificate of the CA.

associateWebsiteCertificateAuthorityResponse_websiteCaId :: Lens' AssociateWebsiteCertificateAuthorityResponse (Maybe Text)

A unique identifier for the CA.

associateWebsiteCertificateAuthorityResponse_httpStatus :: Lens' AssociateWebsiteCertificateAuthorityResponse Int

The response's http status code.

ListDomains

listDomains_nextToken :: Lens' ListDomains (Maybe Text)

The pagination token used to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

listDomains_maxResults :: Lens' ListDomains (Maybe Natural)

The maximum number of results to be included in the next page.

listDomains_fleetArn :: Lens' ListDomains Text

The ARN of the fleet.

listDomainsResponse_nextToken :: Lens' ListDomainsResponse (Maybe Text)

The pagination token used to retrieve the next page of results for this operation. If there are no more pages, this value is null.

listDomainsResponse_domains :: Lens' ListDomainsResponse (Maybe [DomainSummary])

Information about the domains.

listDomainsResponse_httpStatus :: Lens' ListDomainsResponse Int

The response's http status code.

ListDevices

listDevices_nextToken :: Lens' ListDevices (Maybe Text)

The pagination token used to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

listDevices_maxResults :: Lens' ListDevices (Maybe Natural)

The maximum number of results to be included in the next page.

listDevices_fleetArn :: Lens' ListDevices Text

The ARN of the fleet.

listDevicesResponse_nextToken :: Lens' ListDevicesResponse (Maybe Text)

The pagination token used to retrieve the next page of results for this operation. If there are no more pages, this value is null.

listDevicesResponse_devices :: Lens' ListDevicesResponse (Maybe [DeviceSummary])

Information about the devices.

listDevicesResponse_httpStatus :: Lens' ListDevicesResponse Int

The response's http status code.

Types

DeviceSummary

deviceSummary_deviceStatus :: Lens' DeviceSummary (Maybe DeviceStatus)

The status of the device.

deviceSummary_deviceId :: Lens' DeviceSummary (Maybe Text)

The ID of the device.

DomainSummary

domainSummary_displayName :: Lens' DomainSummary (Maybe Text)

The name to display.

domainSummary_domainName :: Lens' DomainSummary Text

The name of the domain.

domainSummary_createdTime :: Lens' DomainSummary UTCTime

The time that the domain was created.

domainSummary_domainStatus :: Lens' DomainSummary DomainStatus

The status of the domain.

FleetSummary

fleetSummary_lastUpdatedTime :: Lens' FleetSummary (Maybe UTCTime)

The time when the fleet was last updated.

fleetSummary_fleetStatus :: Lens' FleetSummary (Maybe FleetStatus)

The status of the fleet.

fleetSummary_companyCode :: Lens' FleetSummary (Maybe Text)

The identifier used by users to sign into the Amazon WorkLink app.

fleetSummary_createdTime :: Lens' FleetSummary (Maybe UTCTime)

The time when the fleet was created.

fleetSummary_fleetArn :: Lens' FleetSummary (Maybe Text)

The Amazon Resource Name (ARN) of the fleet.

fleetSummary_displayName :: Lens' FleetSummary (Maybe Text)

The name of the fleet to display.

fleetSummary_fleetName :: Lens' FleetSummary (Maybe Text)

The name of the fleet.

fleetSummary_tags :: Lens' FleetSummary (Maybe (HashMap Text Text))

The tags attached to the resource. A tag is a key-value pair.

WebsiteAuthorizationProviderSummary

websiteAuthorizationProviderSummary_authorizationProviderId :: Lens' WebsiteAuthorizationProviderSummary (Maybe Text)

A unique identifier for the authorization provider.

websiteAuthorizationProviderSummary_createdTime :: Lens' WebsiteAuthorizationProviderSummary (Maybe UTCTime)

The time of creation.

websiteAuthorizationProviderSummary_domainName :: Lens' WebsiteAuthorizationProviderSummary (Maybe Text)

The domain name of the authorization provider. This applies only to SAML-based authorization providers.

websiteAuthorizationProviderSummary_authorizationProviderType :: Lens' WebsiteAuthorizationProviderSummary AuthorizationProviderType

The authorization provider type.

WebsiteCaSummary

websiteCaSummary_createdTime :: Lens' WebsiteCaSummary (Maybe UTCTime)

The time when the CA was added.

websiteCaSummary_websiteCaId :: Lens' WebsiteCaSummary (Maybe Text)

A unique identifier for the CA.

websiteCaSummary_displayName :: Lens' WebsiteCaSummary (Maybe Text)

The name to display.