Network.AWS.WorkLink.SignOutUser

Description

Signs the user out from all of their devices. The user can sign in again if they have valid credentials.

Creating a Request

data SignOutUser

See: newSignOutUser smart constructor.

Constructors

SignOutUser'

Fields

fleetArn :: Text

The ARN of the fleet.

username :: Text

The name of the user.

newSignOutUser

Arguments

:: Text
$sel:fleetArn:SignOutUser'

-> Text
$sel:username:SignOutUser'

-> SignOutUser

Create a value of SignOutUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetArn:SignOutUser', signOutUser_fleetArn - The ARN of the fleet.

$sel:username:SignOutUser', signOutUser_username - The name of the user.

Request Lenses

signOutUser_fleetArn :: Lens' SignOutUser Text

The ARN of the fleet.

signOutUser_username :: Lens' SignOutUser Text

The name of the user.

Destructuring the Response

data SignOutUserResponse

See: newSignOutUserResponse smart constructor.

Constructors

SignOutUserResponse'

Fields

httpStatus :: Int

The response's http status code.

newSignOutUserResponse

Arguments

:: Int
$sel:httpStatus:SignOutUserResponse'

-> SignOutUserResponse

Create a value of SignOutUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:SignOutUserResponse', signOutUserResponse_httpStatus - The response's http status code.

Response Lenses

signOutUserResponse_httpStatus :: Lens' SignOutUserResponse Int

The response's http status code.