Network.AWS.WorkLink.Types

Description

Service Configuration

defaultService :: Service

API version 2018-09-25 of the Amazon WorkLink SDK configuration.

Errors

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request is not valid.

_ResourceAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The resource already exists.

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

The number of requests exceeds the limit.

_InternalServerErrorException :: AsError a => Getting (First ServiceError) a ServiceError

The service is temporarily unavailable.

_UnauthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

You are not authorized to perform this action.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The requested resource was not found.

AuthorizationProviderType

newtype AuthorizationProviderType

Constructors

AuthorizationProviderType'

Fields

fromAuthorizationProviderType :: Text

Bundled Patterns

pattern AuthorizationProviderType_SAML :: AuthorizationProviderType

DeviceStatus

newtype DeviceStatus

Constructors

DeviceStatus'

Fields

fromDeviceStatus :: Text

Bundled Patterns

pattern DeviceStatus_ACTIVE :: DeviceStatus

pattern DeviceStatus_SIGNED_OUT :: DeviceStatus

DomainStatus

newtype DomainStatus

Constructors

DomainStatus'

Fields

fromDomainStatus :: Text

Bundled Patterns

pattern DomainStatus_ACTIVE :: DomainStatus

pattern DomainStatus_ASSOCIATING :: DomainStatus

pattern DomainStatus_DISASSOCIATED :: DomainStatus

pattern DomainStatus_DISASSOCIATING :: DomainStatus

pattern DomainStatus_FAILED_TO_ASSOCIATE :: DomainStatus

pattern DomainStatus_FAILED_TO_DISASSOCIATE :: DomainStatus

pattern DomainStatus_INACTIVE :: DomainStatus

pattern DomainStatus_PENDING_VALIDATION :: DomainStatus

FleetStatus

newtype FleetStatus

Constructors

FleetStatus'

Fields

fromFleetStatus :: Text

Bundled Patterns

pattern FleetStatus_ACTIVE :: FleetStatus

pattern FleetStatus_CREATING :: FleetStatus

pattern FleetStatus_DELETED :: FleetStatus

pattern FleetStatus_DELETING :: FleetStatus

pattern FleetStatus_FAILED_TO_CREATE :: FleetStatus

pattern FleetStatus_FAILED_TO_DELETE :: FleetStatus

IdentityProviderType

newtype IdentityProviderType

Constructors

IdentityProviderType'

Fields

fromIdentityProviderType :: Text

Bundled Patterns

pattern IdentityProviderType_SAML :: IdentityProviderType

DeviceSummary

data DeviceSummary

The summary of devices.

See: newDeviceSummary smart constructor.

Constructors

DeviceSummary'

Fields

deviceStatus :: Maybe DeviceStatus

The status of the device.

deviceId :: Maybe Text

The ID of the device.

newDeviceSummary :: DeviceSummary

Create a value of DeviceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceStatus:DeviceSummary', deviceSummary_deviceStatus - The status of the device.

$sel:deviceId:DeviceSummary', deviceSummary_deviceId - The ID of the device.

deviceSummary_deviceStatus :: Lens' DeviceSummary (Maybe DeviceStatus)

The status of the device.

deviceSummary_deviceId :: Lens' DeviceSummary (Maybe Text)

The ID of the device.

DomainSummary

data DomainSummary

The summary of the domain.

See: newDomainSummary smart constructor.

Constructors

DomainSummary'

Fields

displayName :: Maybe Text

The name to display.

domainName :: Text

The name of the domain.

createdTime :: POSIX

The time that the domain was created.

domainStatus :: DomainStatus

The status of the domain.

newDomainSummary

Arguments

:: Text
$sel:domainName:DomainSummary'

-> UTCTime
$sel:createdTime:DomainSummary'

-> DomainStatus
$sel:domainStatus:DomainSummary'

-> DomainSummary

Create a value of DomainSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:displayName:DomainSummary', domainSummary_displayName - The name to display.

$sel:domainName:DomainSummary', domainSummary_domainName - The name of the domain.

$sel:createdTime:DomainSummary', domainSummary_createdTime - The time that the domain was created.

$sel:domainStatus:DomainSummary', domainSummary_domainStatus - The status of the domain.

domainSummary_displayName :: Lens' DomainSummary (Maybe Text)

The name to display.

domainSummary_domainName :: Lens' DomainSummary Text

The name of the domain.

domainSummary_createdTime :: Lens' DomainSummary UTCTime

The time that the domain was created.

domainSummary_domainStatus :: Lens' DomainSummary DomainStatus

The status of the domain.

FleetSummary

data FleetSummary

The summary of the fleet.

See: newFleetSummary smart constructor.

Constructors

FleetSummary'

Fields

lastUpdatedTime :: Maybe POSIX

The time when the fleet was last updated.

fleetStatus :: Maybe FleetStatus

The status of the fleet.

companyCode :: Maybe Text

The identifier used by users to sign into the Amazon WorkLink app.

createdTime :: Maybe POSIX

The time when the fleet was created.

fleetArn :: Maybe Text

The Amazon Resource Name (ARN) of the fleet.

displayName :: Maybe Text

The name of the fleet to display.

fleetName :: Maybe Text

The name of the fleet.

tags :: Maybe (HashMap Text Text)

The tags attached to the resource. A tag is a key-value pair.

newFleetSummary :: FleetSummary

Create a value of FleetSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedTime:FleetSummary', fleetSummary_lastUpdatedTime - The time when the fleet was last updated.

$sel:fleetStatus:FleetSummary', fleetSummary_fleetStatus - The status of the fleet.

$sel:companyCode:FleetSummary', fleetSummary_companyCode - The identifier used by users to sign into the Amazon WorkLink app.

$sel:createdTime:FleetSummary', fleetSummary_createdTime - The time when the fleet was created.

$sel:fleetArn:FleetSummary', fleetSummary_fleetArn - The Amazon Resource Name (ARN) of the fleet.

$sel:displayName:FleetSummary', fleetSummary_displayName - The name of the fleet to display.

$sel:fleetName:FleetSummary', fleetSummary_fleetName - The name of the fleet.

$sel:tags:FleetSummary', fleetSummary_tags - The tags attached to the resource. A tag is a key-value pair.

fleetSummary_lastUpdatedTime :: Lens' FleetSummary (Maybe UTCTime)

The time when the fleet was last updated.

fleetSummary_fleetStatus :: Lens' FleetSummary (Maybe FleetStatus)

The status of the fleet.

fleetSummary_companyCode :: Lens' FleetSummary (Maybe Text)

The identifier used by users to sign into the Amazon WorkLink app.

fleetSummary_createdTime :: Lens' FleetSummary (Maybe UTCTime)

The time when the fleet was created.

fleetSummary_fleetArn :: Lens' FleetSummary (Maybe Text)

The Amazon Resource Name (ARN) of the fleet.

fleetSummary_displayName :: Lens' FleetSummary (Maybe Text)

The name of the fleet to display.

fleetSummary_fleetName :: Lens' FleetSummary (Maybe Text)

The name of the fleet.

fleetSummary_tags :: Lens' FleetSummary (Maybe (HashMap Text Text))

The tags attached to the resource. A tag is a key-value pair.

WebsiteAuthorizationProviderSummary

data WebsiteAuthorizationProviderSummary

The summary of the website authorization provider.

See: newWebsiteAuthorizationProviderSummary smart constructor.

Constructors

WebsiteAuthorizationProviderSummary'

Fields

authorizationProviderId :: Maybe Text

A unique identifier for the authorization provider.

createdTime :: Maybe POSIX

The time of creation.

domainName :: Maybe Text

The domain name of the authorization provider. This applies only to SAML-based authorization providers.

authorizationProviderType :: AuthorizationProviderType

The authorization provider type.

newWebsiteAuthorizationProviderSummary

Arguments

:: AuthorizationProviderType
$sel:authorizationProviderType:WebsiteAuthorizationProviderSummary'

-> WebsiteAuthorizationProviderSummary

Create a value of WebsiteAuthorizationProviderSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorizationProviderId:WebsiteAuthorizationProviderSummary', websiteAuthorizationProviderSummary_authorizationProviderId - A unique identifier for the authorization provider.

$sel:createdTime:WebsiteAuthorizationProviderSummary', websiteAuthorizationProviderSummary_createdTime - The time of creation.

$sel:domainName:WebsiteAuthorizationProviderSummary', websiteAuthorizationProviderSummary_domainName - The domain name of the authorization provider. This applies only to SAML-based authorization providers.

$sel:authorizationProviderType:WebsiteAuthorizationProviderSummary', websiteAuthorizationProviderSummary_authorizationProviderType - The authorization provider type.

websiteAuthorizationProviderSummary_authorizationProviderId :: Lens' WebsiteAuthorizationProviderSummary (Maybe Text)

A unique identifier for the authorization provider.

websiteAuthorizationProviderSummary_createdTime :: Lens' WebsiteAuthorizationProviderSummary (Maybe UTCTime)

The time of creation.

websiteAuthorizationProviderSummary_domainName :: Lens' WebsiteAuthorizationProviderSummary (Maybe Text)

The domain name of the authorization provider. This applies only to SAML-based authorization providers.

websiteAuthorizationProviderSummary_authorizationProviderType :: Lens' WebsiteAuthorizationProviderSummary AuthorizationProviderType

The authorization provider type.

WebsiteCaSummary

data WebsiteCaSummary

The summary of the certificate authority (CA).

See: newWebsiteCaSummary smart constructor.

Constructors

WebsiteCaSummary'

Fields

createdTime :: Maybe POSIX

The time when the CA was added.

websiteCaId :: Maybe Text

A unique identifier for the CA.

displayName :: Maybe Text

The name to display.

newWebsiteCaSummary :: WebsiteCaSummary

Create a value of WebsiteCaSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdTime:WebsiteCaSummary', websiteCaSummary_createdTime - The time when the CA was added.

$sel:websiteCaId:WebsiteCaSummary', websiteCaSummary_websiteCaId - A unique identifier for the CA.

$sel:displayName:WebsiteCaSummary', websiteCaSummary_displayName - The name to display.

websiteCaSummary_createdTime :: Lens' WebsiteCaSummary (Maybe UTCTime)

The time when the CA was added.

websiteCaSummary_websiteCaId :: Lens' WebsiteCaSummary (Maybe Text)

A unique identifier for the CA.

websiteCaSummary_displayName :: Lens' WebsiteCaSummary (Maybe Text)

The name to display.