Network.AWS.WorkLink.UpdateCompanyNetworkConfiguration

Description

Updates the company network configuration for the fleet.

Creating a Request

data UpdateCompanyNetworkConfiguration

See: newUpdateCompanyNetworkConfiguration smart constructor.

Constructors

UpdateCompanyNetworkConfiguration'

Fields

fleetArn :: Text

The ARN of the fleet.

vpcId :: Text

The VPC with connectivity to associated websites.

subnetIds :: [Text]

The subnets used for X-ENI connections from Amazon WorkLink rendering containers.

securityGroupIds :: [Text]

The security groups associated with access to the provided subnets.

newUpdateCompanyNetworkConfiguration

Arguments

:: Text
$sel:fleetArn:UpdateCompanyNetworkConfiguration'

-> Text
$sel:vpcId:UpdateCompanyNetworkConfiguration'

-> UpdateCompanyNetworkConfiguration

Create a value of UpdateCompanyNetworkConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleetArn:UpdateCompanyNetworkConfiguration', updateCompanyNetworkConfiguration_fleetArn - The ARN of the fleet.

$sel:vpcId:UpdateCompanyNetworkConfiguration', updateCompanyNetworkConfiguration_vpcId - The VPC with connectivity to associated websites.

$sel:subnetIds:UpdateCompanyNetworkConfiguration', updateCompanyNetworkConfiguration_subnetIds - The subnets used for X-ENI connections from Amazon WorkLink rendering containers.

$sel:securityGroupIds:UpdateCompanyNetworkConfiguration', updateCompanyNetworkConfiguration_securityGroupIds - The security groups associated with access to the provided subnets.

Request Lenses

updateCompanyNetworkConfiguration_fleetArn :: Lens' UpdateCompanyNetworkConfiguration Text

The ARN of the fleet.

updateCompanyNetworkConfiguration_vpcId :: Lens' UpdateCompanyNetworkConfiguration Text

The VPC with connectivity to associated websites.

updateCompanyNetworkConfiguration_subnetIds :: Lens' UpdateCompanyNetworkConfiguration [Text]

The subnets used for X-ENI connections from Amazon WorkLink rendering containers.

updateCompanyNetworkConfiguration_securityGroupIds :: Lens' UpdateCompanyNetworkConfiguration [Text]

The security groups associated with access to the provided subnets.

Destructuring the Response

data UpdateCompanyNetworkConfigurationResponse

See: newUpdateCompanyNetworkConfigurationResponse smart constructor.

Constructors

UpdateCompanyNetworkConfigurationResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateCompanyNetworkConfigurationResponse

Arguments

:: Int
$sel:httpStatus:UpdateCompanyNetworkConfigurationResponse'

-> UpdateCompanyNetworkConfigurationResponse

Create a value of UpdateCompanyNetworkConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateCompanyNetworkConfigurationResponse', updateCompanyNetworkConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

updateCompanyNetworkConfigurationResponse_httpStatus :: Lens' UpdateCompanyNetworkConfigurationResponse Int

The response's http status code.