Network.AWS.WorkMail.AssociateDelegateToResource

Description

Adds a member (user or group) to the resource's set of delegates.

Creating a Request

data AssociateDelegateToResource

See: newAssociateDelegateToResource smart constructor.

Constructors

AssociateDelegateToResource'

Fields

organizationId :: Text

The organization under which the resource exists.

resourceId :: Text

The resource for which members (users or groups) are associated.

entityId :: Text

The member (user or group) to associate to the resource.

newAssociateDelegateToResource

Arguments

:: Text
$sel:organizationId:AssociateDelegateToResource'

-> Text
$sel:resourceId:AssociateDelegateToResource'

-> Text
$sel:entityId:AssociateDelegateToResource'

-> AssociateDelegateToResource

Create a value of AssociateDelegateToResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:organizationId:AssociateDelegateToResource', associateDelegateToResource_organizationId - The organization under which the resource exists.

$sel:resourceId:AssociateDelegateToResource', associateDelegateToResource_resourceId - The resource for which members (users or groups) are associated.

$sel:entityId:AssociateDelegateToResource', associateDelegateToResource_entityId - The member (user or group) to associate to the resource.

Request Lenses

associateDelegateToResource_organizationId :: Lens' AssociateDelegateToResource Text

The organization under which the resource exists.

associateDelegateToResource_resourceId :: Lens' AssociateDelegateToResource Text

The resource for which members (users or groups) are associated.

associateDelegateToResource_entityId :: Lens' AssociateDelegateToResource Text

The member (user or group) to associate to the resource.

Destructuring the Response

data AssociateDelegateToResourceResponse

See: newAssociateDelegateToResourceResponse smart constructor.

Constructors

AssociateDelegateToResourceResponse'

Fields

httpStatus :: Int

The response's http status code.

newAssociateDelegateToResourceResponse

Arguments

:: Int
$sel:httpStatus:AssociateDelegateToResourceResponse'

-> AssociateDelegateToResourceResponse

Create a value of AssociateDelegateToResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AssociateDelegateToResourceResponse', associateDelegateToResourceResponse_httpStatus - The response's http status code.

Response Lenses

associateDelegateToResourceResponse_httpStatus :: Lens' AssociateDelegateToResourceResponse Int

The response's http status code.