Network.AWS.WorkMail.CreateMobileDeviceAccessRule

Description

Creates a new mobile device access rule for the specified Amazon WorkMail organization.

Creating a Request

data CreateMobileDeviceAccessRule

See: newCreateMobileDeviceAccessRule smart constructor.

Constructors

CreateMobileDeviceAccessRule'

Fields

notDeviceModels :: Maybe (NonEmpty Text)

Device models that the rule will not match. All other device models will match.

notDeviceOperatingSystems :: Maybe (NonEmpty Text)

Device operating systems that the rule will not match. All other device operating systems will match.

deviceUserAgents :: Maybe (NonEmpty Text)

Device user agents that the rule will match.

deviceModels :: Maybe (NonEmpty Text)

Device models that the rule will match.

notDeviceTypes :: Maybe (NonEmpty Text)

Device types that the rule will not match. All other device types will match.

description :: Maybe Text

The rule description.

notDeviceUserAgents :: Maybe (NonEmpty Text)

Device user agents that the rule will not match. All other device user agents will match.

deviceTypes :: Maybe (NonEmpty Text)

Device types that the rule will match.

deviceOperatingSystems :: Maybe (NonEmpty Text)

Device operating systems that the rule will match.

clientToken :: Maybe Text

The idempotency token for the client request.

organizationId :: Text

The Amazon WorkMail organization under which the rule will be created.

name :: Text

The rule name.

effect :: MobileDeviceAccessRuleEffect

The effect of the rule when it matches. Allowed values are ALLOW or DENY.

newCreateMobileDeviceAccessRule

Arguments

:: Text
$sel:organizationId:CreateMobileDeviceAccessRule'

-> Text
$sel:name:CreateMobileDeviceAccessRule'

-> MobileDeviceAccessRuleEffect
$sel:effect:CreateMobileDeviceAccessRule'

-> CreateMobileDeviceAccessRule

Create a value of CreateMobileDeviceAccessRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notDeviceModels:CreateMobileDeviceAccessRule', createMobileDeviceAccessRule_notDeviceModels - Device models that the rule will not match. All other device models will match.

$sel:notDeviceOperatingSystems:CreateMobileDeviceAccessRule', createMobileDeviceAccessRule_notDeviceOperatingSystems - Device operating systems that the rule will not match. All other device operating systems will match.

$sel:deviceUserAgents:CreateMobileDeviceAccessRule', createMobileDeviceAccessRule_deviceUserAgents - Device user agents that the rule will match.

$sel:deviceModels:CreateMobileDeviceAccessRule', createMobileDeviceAccessRule_deviceModels - Device models that the rule will match.

$sel:notDeviceTypes:CreateMobileDeviceAccessRule', createMobileDeviceAccessRule_notDeviceTypes - Device types that the rule will not match. All other device types will match.

$sel:description:CreateMobileDeviceAccessRule', createMobileDeviceAccessRule_description - The rule description.

$sel:notDeviceUserAgents:CreateMobileDeviceAccessRule', createMobileDeviceAccessRule_notDeviceUserAgents - Device user agents that the rule will not match. All other device user agents will match.

$sel:deviceTypes:CreateMobileDeviceAccessRule', createMobileDeviceAccessRule_deviceTypes - Device types that the rule will match.

$sel:deviceOperatingSystems:CreateMobileDeviceAccessRule', createMobileDeviceAccessRule_deviceOperatingSystems - Device operating systems that the rule will match.

$sel:clientToken:CreateMobileDeviceAccessRule', createMobileDeviceAccessRule_clientToken - The idempotency token for the client request.

$sel:organizationId:CreateMobileDeviceAccessRule', createMobileDeviceAccessRule_organizationId - The Amazon WorkMail organization under which the rule will be created.

$sel:name:CreateMobileDeviceAccessRule', createMobileDeviceAccessRule_name - The rule name.

$sel:effect:CreateMobileDeviceAccessRule', createMobileDeviceAccessRule_effect - The effect of the rule when it matches. Allowed values are ALLOW or DENY.

Request Lenses

createMobileDeviceAccessRule_notDeviceModels :: Lens' CreateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device models that the rule will not match. All other device models will match.

createMobileDeviceAccessRule_notDeviceOperatingSystems :: Lens' CreateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device operating systems that the rule will not match. All other device operating systems will match.

createMobileDeviceAccessRule_deviceUserAgents :: Lens' CreateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device user agents that the rule will match.

createMobileDeviceAccessRule_deviceModels :: Lens' CreateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device models that the rule will match.

createMobileDeviceAccessRule_notDeviceTypes :: Lens' CreateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device types that the rule will not match. All other device types will match.

createMobileDeviceAccessRule_description :: Lens' CreateMobileDeviceAccessRule (Maybe Text)

The rule description.

createMobileDeviceAccessRule_notDeviceUserAgents :: Lens' CreateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device user agents that the rule will not match. All other device user agents will match.

createMobileDeviceAccessRule_deviceTypes :: Lens' CreateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device types that the rule will match.

createMobileDeviceAccessRule_deviceOperatingSystems :: Lens' CreateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device operating systems that the rule will match.

createMobileDeviceAccessRule_clientToken :: Lens' CreateMobileDeviceAccessRule (Maybe Text)

The idempotency token for the client request.

createMobileDeviceAccessRule_organizationId :: Lens' CreateMobileDeviceAccessRule Text

The Amazon WorkMail organization under which the rule will be created.

createMobileDeviceAccessRule_name :: Lens' CreateMobileDeviceAccessRule Text

The rule name.

createMobileDeviceAccessRule_effect :: Lens' CreateMobileDeviceAccessRule MobileDeviceAccessRuleEffect

The effect of the rule when it matches. Allowed values are ALLOW or DENY.

Destructuring the Response

data CreateMobileDeviceAccessRuleResponse

See: newCreateMobileDeviceAccessRuleResponse smart constructor.

Constructors

CreateMobileDeviceAccessRuleResponse'

Fields

mobileDeviceAccessRuleId :: Maybe Text

The identifier for the newly created mobile device access rule.

httpStatus :: Int

The response's http status code.

newCreateMobileDeviceAccessRuleResponse

Arguments

:: Int
$sel:httpStatus:CreateMobileDeviceAccessRuleResponse'

-> CreateMobileDeviceAccessRuleResponse

Create a value of CreateMobileDeviceAccessRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mobileDeviceAccessRuleId:CreateMobileDeviceAccessRuleResponse', createMobileDeviceAccessRuleResponse_mobileDeviceAccessRuleId - The identifier for the newly created mobile device access rule.

$sel:httpStatus:CreateMobileDeviceAccessRuleResponse', createMobileDeviceAccessRuleResponse_httpStatus - The response's http status code.

Response Lenses

createMobileDeviceAccessRuleResponse_mobileDeviceAccessRuleId :: Lens' CreateMobileDeviceAccessRuleResponse (Maybe Text)

The identifier for the newly created mobile device access rule.

createMobileDeviceAccessRuleResponse_httpStatus :: Lens' CreateMobileDeviceAccessRuleResponse Int

The response's http status code.