Network.AWS.WorkMail.DeleteOrganization

Description

Deletes an Amazon WorkMail organization and all underlying AWS resources managed by Amazon WorkMail as part of the organization. You can choose whether to delete the associated directory. For more information, see Removing an organization in the Amazon WorkMail Administrator Guide.

Creating a Request

data DeleteOrganization

See: newDeleteOrganization smart constructor.

Constructors

DeleteOrganization'

Fields

clientToken :: Maybe Text

The idempotency token associated with the request.

organizationId :: Text

The organization ID.

deleteDirectory :: Bool

If true, deletes the AWS Directory Service directory associated with the organization.

newDeleteOrganization

Arguments

:: Text
$sel:organizationId:DeleteOrganization'

-> Bool
$sel:deleteDirectory:DeleteOrganization'

-> DeleteOrganization

Create a value of DeleteOrganization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:DeleteOrganization', deleteOrganization_clientToken - The idempotency token associated with the request.

$sel:organizationId:DeleteOrganization', deleteOrganization_organizationId - The organization ID.

$sel:deleteDirectory:DeleteOrganization', deleteOrganization_deleteDirectory - If true, deletes the AWS Directory Service directory associated with the organization.

Request Lenses

deleteOrganization_clientToken :: Lens' DeleteOrganization (Maybe Text)

The idempotency token associated with the request.

deleteOrganization_organizationId :: Lens' DeleteOrganization Text

The organization ID.

deleteOrganization_deleteDirectory :: Lens' DeleteOrganization Bool

If true, deletes the AWS Directory Service directory associated with the organization.

Destructuring the Response

data DeleteOrganizationResponse

See: newDeleteOrganizationResponse smart constructor.

Constructors

DeleteOrganizationResponse'

Fields

organizationId :: Maybe Text

The organization ID.

state :: Maybe Text

The state of the organization.

httpStatus :: Int

The response's http status code.

newDeleteOrganizationResponse

Arguments

:: Int
$sel:httpStatus:DeleteOrganizationResponse'

-> DeleteOrganizationResponse

Create a value of DeleteOrganizationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:organizationId:DeleteOrganization', deleteOrganizationResponse_organizationId - The organization ID.

$sel:state:DeleteOrganizationResponse', deleteOrganizationResponse_state - The state of the organization.

$sel:httpStatus:DeleteOrganizationResponse', deleteOrganizationResponse_httpStatus - The response's http status code.

Response Lenses

deleteOrganizationResponse_organizationId :: Lens' DeleteOrganizationResponse (Maybe Text)

The organization ID.

deleteOrganizationResponse_state :: Lens' DeleteOrganizationResponse (Maybe Text)

The state of the organization.

deleteOrganizationResponse_httpStatus :: Lens' DeleteOrganizationResponse Int

The response's http status code.