Network.AWS.WorkMail.DescribeOrganization

Description

Provides more information regarding a given organization based on its identifier.

Creating a Request

data DescribeOrganization

See: newDescribeOrganization smart constructor.

Constructors

DescribeOrganization'

Fields

organizationId :: Text

The identifier for the organization to be described.

newDescribeOrganization

Arguments

:: Text
$sel:organizationId:DescribeOrganization'

-> DescribeOrganization

Create a value of DescribeOrganization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:organizationId:DescribeOrganization', describeOrganization_organizationId - The identifier for the organization to be described.

Request Lenses

describeOrganization_organizationId :: Lens' DescribeOrganization Text

The identifier for the organization to be described.

Destructuring the Response

data DescribeOrganizationResponse

See: newDescribeOrganizationResponse smart constructor.

Constructors

DescribeOrganizationResponse'

Fields

directoryId :: Maybe Text

The identifier for the directory associated with an Amazon WorkMail organization.

state :: Maybe Text

The state of an organization.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the organization.

alias :: Maybe Text

The alias for an organization.

completedDate :: Maybe POSIX

The date at which the organization became usable in the WorkMail context, in UNIX epoch time format.

directoryType :: Maybe Text

The type of directory associated with the WorkMail organization.

defaultMailDomain :: Maybe Text

The default mail domain associated with the organization.

errorMessage :: Maybe Text

(Optional) The error message indicating if unexpected behavior was encountered with regards to the organization.

organizationId :: Maybe Text

The identifier of an organization.

httpStatus :: Int

The response's http status code.

newDescribeOrganizationResponse

Arguments

:: Int
$sel:httpStatus:DescribeOrganizationResponse'

-> DescribeOrganizationResponse

Create a value of DescribeOrganizationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryId:DescribeOrganizationResponse', describeOrganizationResponse_directoryId - The identifier for the directory associated with an Amazon WorkMail organization.

$sel:state:DescribeOrganizationResponse', describeOrganizationResponse_state - The state of an organization.

$sel:arn:DescribeOrganizationResponse', describeOrganizationResponse_arn - The Amazon Resource Name (ARN) of the organization.

$sel:alias:DescribeOrganizationResponse', describeOrganizationResponse_alias - The alias for an organization.

$sel:completedDate:DescribeOrganizationResponse', describeOrganizationResponse_completedDate - The date at which the organization became usable in the WorkMail context, in UNIX epoch time format.

$sel:directoryType:DescribeOrganizationResponse', describeOrganizationResponse_directoryType - The type of directory associated with the WorkMail organization.

$sel:defaultMailDomain:DescribeOrganizationResponse', describeOrganizationResponse_defaultMailDomain - The default mail domain associated with the organization.

$sel:errorMessage:DescribeOrganizationResponse', describeOrganizationResponse_errorMessage - (Optional) The error message indicating if unexpected behavior was encountered with regards to the organization.

$sel:organizationId:DescribeOrganization', describeOrganizationResponse_organizationId - The identifier of an organization.

$sel:httpStatus:DescribeOrganizationResponse', describeOrganizationResponse_httpStatus - The response's http status code.

Response Lenses

describeOrganizationResponse_directoryId :: Lens' DescribeOrganizationResponse (Maybe Text)

The identifier for the directory associated with an Amazon WorkMail organization.

describeOrganizationResponse_state :: Lens' DescribeOrganizationResponse (Maybe Text)

The state of an organization.

describeOrganizationResponse_arn :: Lens' DescribeOrganizationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the organization.

describeOrganizationResponse_alias :: Lens' DescribeOrganizationResponse (Maybe Text)

The alias for an organization.

describeOrganizationResponse_completedDate :: Lens' DescribeOrganizationResponse (Maybe UTCTime)

The date at which the organization became usable in the WorkMail context, in UNIX epoch time format.

describeOrganizationResponse_directoryType :: Lens' DescribeOrganizationResponse (Maybe Text)

The type of directory associated with the WorkMail organization.

describeOrganizationResponse_defaultMailDomain :: Lens' DescribeOrganizationResponse (Maybe Text)

The default mail domain associated with the organization.

describeOrganizationResponse_errorMessage :: Lens' DescribeOrganizationResponse (Maybe Text)

(Optional) The error message indicating if unexpected behavior was encountered with regards to the organization.

describeOrganizationResponse_organizationId :: Lens' DescribeOrganizationResponse (Maybe Text)

The identifier of an organization.

describeOrganizationResponse_httpStatus :: Lens' DescribeOrganizationResponse Int

The response's http status code.