Network.AWS.WorkMail.DisassociateDelegateFromResource

Description

Removes a member from the resource's set of delegates.

Creating a Request

data DisassociateDelegateFromResource

See: newDisassociateDelegateFromResource smart constructor.

Constructors

DisassociateDelegateFromResource'

Fields

organizationId :: Text

The identifier for the organization under which the resource exists.

resourceId :: Text

The identifier of the resource from which delegates' set members are removed.

entityId :: Text

The identifier for the member (user, group) to be removed from the resource's delegates.

newDisassociateDelegateFromResource

Arguments

:: Text
$sel:organizationId:DisassociateDelegateFromResource'

-> Text
$sel:resourceId:DisassociateDelegateFromResource'

-> Text
$sel:entityId:DisassociateDelegateFromResource'

-> DisassociateDelegateFromResource

Create a value of DisassociateDelegateFromResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:organizationId:DisassociateDelegateFromResource', disassociateDelegateFromResource_organizationId - The identifier for the organization under which the resource exists.

$sel:resourceId:DisassociateDelegateFromResource', disassociateDelegateFromResource_resourceId - The identifier of the resource from which delegates' set members are removed.

$sel:entityId:DisassociateDelegateFromResource', disassociateDelegateFromResource_entityId - The identifier for the member (user, group) to be removed from the resource's delegates.

Request Lenses

disassociateDelegateFromResource_organizationId :: Lens' DisassociateDelegateFromResource Text

The identifier for the organization under which the resource exists.

disassociateDelegateFromResource_resourceId :: Lens' DisassociateDelegateFromResource Text

The identifier of the resource from which delegates' set members are removed.

disassociateDelegateFromResource_entityId :: Lens' DisassociateDelegateFromResource Text

The identifier for the member (user, group) to be removed from the resource's delegates.

Destructuring the Response

data DisassociateDelegateFromResourceResponse

See: newDisassociateDelegateFromResourceResponse smart constructor.

Constructors

DisassociateDelegateFromResourceResponse'

Fields

httpStatus :: Int

The response's http status code.

newDisassociateDelegateFromResourceResponse

Arguments

:: Int
$sel:httpStatus:DisassociateDelegateFromResourceResponse'

-> DisassociateDelegateFromResourceResponse

Create a value of DisassociateDelegateFromResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisassociateDelegateFromResourceResponse', disassociateDelegateFromResourceResponse_httpStatus - The response's http status code.

Response Lenses

disassociateDelegateFromResourceResponse_httpStatus :: Lens' DisassociateDelegateFromResourceResponse Int

The response's http status code.