Network.AWS.WorkMail.GetDefaultRetentionPolicy

Description

Gets the default retention policy details for the specified organization.

Creating a Request

data GetDefaultRetentionPolicy

See: newGetDefaultRetentionPolicy smart constructor.

Constructors

GetDefaultRetentionPolicy'

Fields

organizationId :: Text

The organization ID.

newGetDefaultRetentionPolicy

Arguments

:: Text
$sel:organizationId:GetDefaultRetentionPolicy'

-> GetDefaultRetentionPolicy

Create a value of GetDefaultRetentionPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:organizationId:GetDefaultRetentionPolicy', getDefaultRetentionPolicy_organizationId - The organization ID.

Request Lenses

getDefaultRetentionPolicy_organizationId :: Lens' GetDefaultRetentionPolicy Text

The organization ID.

Destructuring the Response

data GetDefaultRetentionPolicyResponse

See: newGetDefaultRetentionPolicyResponse smart constructor.

Constructors

GetDefaultRetentionPolicyResponse'

Fields

id :: Maybe Text

The retention policy ID.

folderConfigurations :: Maybe [FolderConfiguration]

The retention policy folder configurations.

name :: Maybe Text

The retention policy name.

description :: Maybe Text

The retention policy description.

httpStatus :: Int

The response's http status code.

newGetDefaultRetentionPolicyResponse

Arguments

:: Int
$sel:httpStatus:GetDefaultRetentionPolicyResponse'

-> GetDefaultRetentionPolicyResponse

Create a value of GetDefaultRetentionPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:GetDefaultRetentionPolicyResponse', getDefaultRetentionPolicyResponse_id - The retention policy ID.

$sel:folderConfigurations:GetDefaultRetentionPolicyResponse', getDefaultRetentionPolicyResponse_folderConfigurations - The retention policy folder configurations.

$sel:name:GetDefaultRetentionPolicyResponse', getDefaultRetentionPolicyResponse_name - The retention policy name.

$sel:description:GetDefaultRetentionPolicyResponse', getDefaultRetentionPolicyResponse_description - The retention policy description.

$sel:httpStatus:GetDefaultRetentionPolicyResponse', getDefaultRetentionPolicyResponse_httpStatus - The response's http status code.

Response Lenses

getDefaultRetentionPolicyResponse_id :: Lens' GetDefaultRetentionPolicyResponse (Maybe Text)

The retention policy ID.

getDefaultRetentionPolicyResponse_folderConfigurations :: Lens' GetDefaultRetentionPolicyResponse (Maybe [FolderConfiguration])

The retention policy folder configurations.

getDefaultRetentionPolicyResponse_name :: Lens' GetDefaultRetentionPolicyResponse (Maybe Text)

The retention policy name.

getDefaultRetentionPolicyResponse_description :: Lens' GetDefaultRetentionPolicyResponse (Maybe Text)

The retention policy description.

getDefaultRetentionPolicyResponse_httpStatus :: Lens' GetDefaultRetentionPolicyResponse Int

The response's http status code.