Network.AWS.WorkMail.GetMailDomain

Description

Gets details for a mail domain, including domain records required to configure your domain with recommended security.

Creating a Request

data GetMailDomain

See: newGetMailDomain smart constructor.

Constructors

GetMailDomain'

Fields

organizationId :: Text

The Amazon WorkMail organization for which the domain is retrieved.

domainName :: Text

The domain from which you want to retrieve details.

newGetMailDomain

Arguments

:: Text
$sel:organizationId:GetMailDomain'

-> Text
$sel:domainName:GetMailDomain'

-> GetMailDomain

Create a value of GetMailDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:organizationId:GetMailDomain', getMailDomain_organizationId - The Amazon WorkMail organization for which the domain is retrieved.

$sel:domainName:GetMailDomain', getMailDomain_domainName - The domain from which you want to retrieve details.

Request Lenses

getMailDomain_organizationId :: Lens' GetMailDomain Text

The Amazon WorkMail organization for which the domain is retrieved.

getMailDomain_domainName :: Lens' GetMailDomain Text

The domain from which you want to retrieve details.

Destructuring the Response

data GetMailDomainResponse

See: newGetMailDomainResponse smart constructor.

Constructors

GetMailDomainResponse'

Fields

isTestDomain :: Maybe Bool

Specifies whether the domain is a test domain provided by WorkMail, or a custom domain.

records :: Maybe [DnsRecord]

A list of the DNS records that Amazon WorkMail recommends adding in your DNS provider for the best user experience. The records configure your domain with DMARC, SPF, DKIM, and direct incoming email traffic to SES. See admin guide for more details.

ownershipVerificationStatus :: Maybe DnsRecordVerificationStatus

Indicates the status of the domain ownership verification.

dkimVerificationStatus :: Maybe DnsRecordVerificationStatus

Indicates the status of a DKIM verification.

isDefault :: Maybe Bool

Specifies whether the domain is the default domain for your organization.

httpStatus :: Int

The response's http status code.

newGetMailDomainResponse

Arguments

:: Int
$sel:httpStatus:GetMailDomainResponse'

-> GetMailDomainResponse

Create a value of GetMailDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isTestDomain:GetMailDomainResponse', getMailDomainResponse_isTestDomain - Specifies whether the domain is a test domain provided by WorkMail, or a custom domain.

$sel:records:GetMailDomainResponse', getMailDomainResponse_records - A list of the DNS records that Amazon WorkMail recommends adding in your DNS provider for the best user experience. The records configure your domain with DMARC, SPF, DKIM, and direct incoming email traffic to SES. See admin guide for more details.

$sel:ownershipVerificationStatus:GetMailDomainResponse', getMailDomainResponse_ownershipVerificationStatus - Indicates the status of the domain ownership verification.

$sel:dkimVerificationStatus:GetMailDomainResponse', getMailDomainResponse_dkimVerificationStatus - Indicates the status of a DKIM verification.

$sel:isDefault:GetMailDomainResponse', getMailDomainResponse_isDefault - Specifies whether the domain is the default domain for your organization.

$sel:httpStatus:GetMailDomainResponse', getMailDomainResponse_httpStatus - The response's http status code.

Response Lenses

getMailDomainResponse_isTestDomain :: Lens' GetMailDomainResponse (Maybe Bool)

Specifies whether the domain is a test domain provided by WorkMail, or a custom domain.

getMailDomainResponse_records :: Lens' GetMailDomainResponse (Maybe [DnsRecord])

A list of the DNS records that Amazon WorkMail recommends adding in your DNS provider for the best user experience. The records configure your domain with DMARC, SPF, DKIM, and direct incoming email traffic to SES. See admin guide for more details.

getMailDomainResponse_ownershipVerificationStatus :: Lens' GetMailDomainResponse (Maybe DnsRecordVerificationStatus)

Indicates the status of the domain ownership verification.

getMailDomainResponse_dkimVerificationStatus :: Lens' GetMailDomainResponse (Maybe DnsRecordVerificationStatus)

Indicates the status of a DKIM verification.

getMailDomainResponse_isDefault :: Lens' GetMailDomainResponse (Maybe Bool)

Specifies whether the domain is the default domain for your organization.

getMailDomainResponse_httpStatus :: Lens' GetMailDomainResponse Int

The response's http status code.