Network.AWS.WorkMail.Lens

Description

Operations

DescribeResource

describeResource_organizationId :: Lens' DescribeResource Text

The identifier associated with the organization for which the resource is described.

describeResource_resourceId :: Lens' DescribeResource Text

The identifier of the resource to be described.

describeResourceResponse_resourceId :: Lens' DescribeResourceResponse (Maybe Text)

The identifier of the described resource.

describeResourceResponse_enabledDate :: Lens' DescribeResourceResponse (Maybe UTCTime)

The date and time when a resource was enabled for WorkMail, in UNIX epoch time format.

describeResourceResponse_name :: Lens' DescribeResourceResponse (Maybe Text)

The name of the described resource.

describeResourceResponse_state :: Lens' DescribeResourceResponse (Maybe EntityState)

The state of the resource: enabled (registered to Amazon WorkMail), disabled (deregistered or never registered to WorkMail), or deleted.

describeResourceResponse_disabledDate :: Lens' DescribeResourceResponse (Maybe UTCTime)

The date and time when a resource was disabled from WorkMail, in UNIX epoch time format.

describeResourceResponse_email :: Lens' DescribeResourceResponse (Maybe Text)

The email of the described resource.

describeResourceResponse_bookingOptions :: Lens' DescribeResourceResponse (Maybe BookingOptions)

The booking options for the described resource.

describeResourceResponse_type :: Lens' DescribeResourceResponse (Maybe ResourceType)

The type of the described resource.

describeResourceResponse_httpStatus :: Lens' DescribeResourceResponse Int

The response's http status code.

DeleteAlias

deleteAlias_organizationId :: Lens' DeleteAlias Text

The identifier for the organization under which the user exists.

deleteAlias_entityId :: Lens' DeleteAlias Text

The identifier for the member (user or group) from which to have the aliases removed.

deleteAlias_alias :: Lens' DeleteAlias Text

The aliases to be removed from the user's set of aliases. Duplicate entries in the list are collapsed into single entries (the list is transformed into a set).

deleteAliasResponse_httpStatus :: Lens' DeleteAliasResponse Int

The response's http status code.

CreateOrganization

createOrganization_enableInteroperability :: Lens' CreateOrganization (Maybe Bool)

When true, allows organization interoperability between Amazon WorkMail and Microsoft Exchange. Can only be set to true if an AD Connector directory ID is included in the request.

createOrganization_domains :: Lens' CreateOrganization (Maybe [Domain])

The email domains to associate with the organization.

createOrganization_kmsKeyArn :: Lens' CreateOrganization (Maybe Text)

The Amazon Resource Name (ARN) of a customer managed master key from AWS KMS.

createOrganization_directoryId :: Lens' CreateOrganization (Maybe Text)

The AWS Directory Service directory ID.

createOrganization_clientToken :: Lens' CreateOrganization (Maybe Text)

The idempotency token associated with the request.

createOrganization_alias :: Lens' CreateOrganization Text

The organization alias.

createOrganizationResponse_organizationId :: Lens' CreateOrganizationResponse (Maybe Text)

The organization ID.

createOrganizationResponse_httpStatus :: Lens' CreateOrganizationResponse Int

The response's http status code.

StartMailboxExportJob

startMailboxExportJob_description :: Lens' StartMailboxExportJob (Maybe Text)

The mailbox export job description.

startMailboxExportJob_clientToken :: Lens' StartMailboxExportJob Text

The idempotency token for the client request.

startMailboxExportJob_organizationId :: Lens' StartMailboxExportJob Text

The identifier associated with the organization.

startMailboxExportJob_entityId :: Lens' StartMailboxExportJob Text

The identifier of the user or resource associated with the mailbox.

startMailboxExportJob_roleArn :: Lens' StartMailboxExportJob Text

The ARN of the AWS Identity and Access Management (IAM) role that grants write permission to the S3 bucket.

startMailboxExportJob_kmsKeyArn :: Lens' StartMailboxExportJob Text

The Amazon Resource Name (ARN) of the symmetric AWS Key Management Service (AWS KMS) key that encrypts the exported mailbox content.

startMailboxExportJob_s3BucketName :: Lens' StartMailboxExportJob Text

The name of the S3 bucket.

startMailboxExportJob_s3Prefix :: Lens' StartMailboxExportJob Text

The S3 bucket prefix.

startMailboxExportJobResponse_jobId :: Lens' StartMailboxExportJobResponse (Maybe Text)

The job ID.

startMailboxExportJobResponse_httpStatus :: Lens' StartMailboxExportJobResponse Int

The response's http status code.

DeleteAccessControlRule

deleteAccessControlRule_organizationId :: Lens' DeleteAccessControlRule Text

The identifier for the organization.

deleteAccessControlRule_name :: Lens' DeleteAccessControlRule Text

The name of the access control rule.

deleteAccessControlRuleResponse_httpStatus :: Lens' DeleteAccessControlRuleResponse Int

The response's http status code.

ListResourceDelegates

listResourceDelegates_nextToken :: Lens' ListResourceDelegates (Maybe Text)

The token used to paginate through the delegates associated with a resource.

listResourceDelegates_maxResults :: Lens' ListResourceDelegates (Maybe Natural)

The number of maximum results in a page.

listResourceDelegates_organizationId :: Lens' ListResourceDelegates Text

The identifier for the organization that contains the resource for which delegates are listed.

listResourceDelegates_resourceId :: Lens' ListResourceDelegates Text

The identifier for the resource whose delegates are listed.

listResourceDelegatesResponse_nextToken :: Lens' ListResourceDelegatesResponse (Maybe Text)

The token used to paginate through the delegates associated with a resource. While results are still available, it has an associated value. When the last page is reached, the token is empty.

listResourceDelegatesResponse_delegates :: Lens' ListResourceDelegatesResponse (Maybe [Delegate])

One page of the resource's delegates.

listResourceDelegatesResponse_httpStatus :: Lens' ListResourceDelegatesResponse Int

The response's http status code.

DisassociateDelegateFromResource

disassociateDelegateFromResource_organizationId :: Lens' DisassociateDelegateFromResource Text

The identifier for the organization under which the resource exists.

disassociateDelegateFromResource_resourceId :: Lens' DisassociateDelegateFromResource Text

The identifier of the resource from which delegates' set members are removed.

disassociateDelegateFromResource_entityId :: Lens' DisassociateDelegateFromResource Text

The identifier for the member (user, group) to be removed from the resource's delegates.

disassociateDelegateFromResourceResponse_httpStatus :: Lens' DisassociateDelegateFromResourceResponse Int

The response's http status code.

GetDefaultRetentionPolicy

getDefaultRetentionPolicy_organizationId :: Lens' GetDefaultRetentionPolicy Text

The organization ID.

getDefaultRetentionPolicyResponse_id :: Lens' GetDefaultRetentionPolicyResponse (Maybe Text)

The retention policy ID.

getDefaultRetentionPolicyResponse_folderConfigurations :: Lens' GetDefaultRetentionPolicyResponse (Maybe [FolderConfiguration])

The retention policy folder configurations.

getDefaultRetentionPolicyResponse_name :: Lens' GetDefaultRetentionPolicyResponse (Maybe Text)

The retention policy name.

getDefaultRetentionPolicyResponse_description :: Lens' GetDefaultRetentionPolicyResponse (Maybe Text)

The retention policy description.

getDefaultRetentionPolicyResponse_httpStatus :: Lens' GetDefaultRetentionPolicyResponse Int

The response's http status code.

ListGroups

listGroups_nextToken :: Lens' ListGroups (Maybe Text)

The token to use to retrieve the next page of results. The first call does not contain any tokens.

listGroups_maxResults :: Lens' ListGroups (Maybe Natural)

The maximum number of results to return in a single call.

listGroups_organizationId :: Lens' ListGroups Text

The identifier for the organization under which the groups exist.

listGroupsResponse_groups :: Lens' ListGroupsResponse (Maybe [Group])

The overview of groups for an organization.

listGroupsResponse_nextToken :: Lens' ListGroupsResponse (Maybe Text)

The token to use to retrieve the next page of results. The value is "null" when there are no more results to return.

listGroupsResponse_httpStatus :: Lens' ListGroupsResponse Int

The response's http status code.

CreateResource

createResource_organizationId :: Lens' CreateResource Text

The identifier associated with the organization for which the resource is created.

createResource_name :: Lens' CreateResource Text

The name of the new resource.

createResource_type :: Lens' CreateResource ResourceType

The type of the new resource. The available types are equipment and room.

createResourceResponse_resourceId :: Lens' CreateResourceResponse (Maybe Text)

The identifier of the new resource.

createResourceResponse_httpStatus :: Lens' CreateResourceResponse Int

The response's http status code.

ListMailboxExportJobs

listMailboxExportJobs_nextToken :: Lens' ListMailboxExportJobs (Maybe Text)

The token to use to retrieve the next page of results.

listMailboxExportJobs_maxResults :: Lens' ListMailboxExportJobs (Maybe Natural)

The maximum number of results to return in a single call.

listMailboxExportJobs_organizationId :: Lens' ListMailboxExportJobs Text

The organization ID.

listMailboxExportJobsResponse_nextToken :: Lens' ListMailboxExportJobsResponse (Maybe Text)

The token to use to retrieve the next page of results.

listMailboxExportJobsResponse_jobs :: Lens' ListMailboxExportJobsResponse (Maybe [MailboxExportJob])

The mailbox export job details.

listMailboxExportJobsResponse_httpStatus :: Lens' ListMailboxExportJobsResponse Int

The response's http status code.

DescribeOrganization

describeOrganization_organizationId :: Lens' DescribeOrganization Text

The identifier for the organization to be described.

describeOrganizationResponse_alias :: Lens' DescribeOrganizationResponse (Maybe Text)

The alias for an organization.

describeOrganizationResponse_organizationId :: Lens' DescribeOrganizationResponse (Maybe Text)

The identifier of an organization.

describeOrganizationResponse_arn :: Lens' DescribeOrganizationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the organization.

describeOrganizationResponse_defaultMailDomain :: Lens' DescribeOrganizationResponse (Maybe Text)

The default mail domain associated with the organization.

describeOrganizationResponse_state :: Lens' DescribeOrganizationResponse (Maybe Text)

The state of an organization.

describeOrganizationResponse_directoryId :: Lens' DescribeOrganizationResponse (Maybe Text)

The identifier for the directory associated with an Amazon WorkMail organization.

describeOrganizationResponse_directoryType :: Lens' DescribeOrganizationResponse (Maybe Text)

The type of directory associated with the WorkMail organization.

describeOrganizationResponse_completedDate :: Lens' DescribeOrganizationResponse (Maybe UTCTime)

The date at which the organization became usable in the WorkMail context, in UNIX epoch time format.

describeOrganizationResponse_errorMessage :: Lens' DescribeOrganizationResponse (Maybe Text)

(Optional) The error message indicating if unexpected behavior was encountered with regards to the organization.

describeOrganizationResponse_httpStatus :: Lens' DescribeOrganizationResponse Int

The response's http status code.

UpdateResource

updateResource_name :: Lens' UpdateResource (Maybe Text)

The name of the resource to be updated.

updateResource_bookingOptions :: Lens' UpdateResource (Maybe BookingOptions)

The resource's booking options to be updated.

updateResource_organizationId :: Lens' UpdateResource Text

The identifier associated with the organization for which the resource is updated.

updateResource_resourceId :: Lens' UpdateResource Text

The identifier of the resource to be updated.

updateResourceResponse_httpStatus :: Lens' UpdateResourceResponse Int

The response's http status code.

DeleteResource

deleteResource_organizationId :: Lens' DeleteResource Text

The identifier associated with the organization from which the resource is deleted.

deleteResource_resourceId :: Lens' DeleteResource Text

The identifier of the resource to be deleted.

deleteResourceResponse_httpStatus :: Lens' DeleteResourceResponse Int

The response's http status code.

ListMobileDeviceAccessRules

listMobileDeviceAccessRules_organizationId :: Lens' ListMobileDeviceAccessRules Text

The Amazon WorkMail organization for which to list the rules.

listMobileDeviceAccessRulesResponse_rules :: Lens' ListMobileDeviceAccessRulesResponse (Maybe [MobileDeviceAccessRule])

The list of mobile device access rules that exist under the specified Amazon WorkMail organization.

listMobileDeviceAccessRulesResponse_httpStatus :: Lens' ListMobileDeviceAccessRulesResponse Int

The response's http status code.

CreateGroup

createGroup_organizationId :: Lens' CreateGroup Text

The organization under which the group is to be created.

createGroup_name :: Lens' CreateGroup Text

The name of the group.

createGroupResponse_groupId :: Lens' CreateGroupResponse (Maybe Text)

The identifier of the group.

createGroupResponse_httpStatus :: Lens' CreateGroupResponse Int

The response's http status code.

UntagResource

untagResource_resourceARN :: Lens' UntagResource Text

The resource ARN.

untagResource_tagKeys :: Lens' UntagResource [Text]

The tag keys.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

TagResource

tagResource_resourceARN :: Lens' TagResource Text

The resource ARN.

tagResource_tags :: Lens' TagResource [Tag]

The tag key-value pairs.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

AssociateMemberToGroup

associateMemberToGroup_organizationId :: Lens' AssociateMemberToGroup Text

The organization under which the group exists.

associateMemberToGroup_groupId :: Lens' AssociateMemberToGroup Text

The group to which the member (user or group) is associated.

associateMemberToGroup_memberId :: Lens' AssociateMemberToGroup Text

The member (user or group) to associate to the group.

associateMemberToGroupResponse_httpStatus :: Lens' AssociateMemberToGroupResponse Int

The response's http status code.

PutRetentionPolicy

putRetentionPolicy_id :: Lens' PutRetentionPolicy (Maybe Text)

The retention policy ID.

putRetentionPolicy_description :: Lens' PutRetentionPolicy (Maybe Text)

The retention policy description.

putRetentionPolicy_organizationId :: Lens' PutRetentionPolicy Text

The organization ID.

putRetentionPolicy_name :: Lens' PutRetentionPolicy Text

The retention policy name.

putRetentionPolicy_folderConfigurations :: Lens' PutRetentionPolicy [FolderConfiguration]

The retention policy folder configurations.

putRetentionPolicyResponse_httpStatus :: Lens' PutRetentionPolicyResponse Int

The response's http status code.

CreateUser

createUser_organizationId :: Lens' CreateUser Text

The identifier of the organization for which the user is created.

createUser_name :: Lens' CreateUser Text

The name for the new user. WorkMail directory user names have a maximum length of 64. All others have a maximum length of 20.

createUser_displayName :: Lens' CreateUser Text

The display name for the new user.

createUser_password :: Lens' CreateUser Text

The password for the new user.

createUserResponse_userId :: Lens' CreateUserResponse (Maybe Text)

The identifier for the new user.

createUserResponse_httpStatus :: Lens' CreateUserResponse Int

The response's http status code.

PutMailboxPermissions

putMailboxPermissions_organizationId :: Lens' PutMailboxPermissions Text

The identifier of the organization under which the user, group, or resource exists.

putMailboxPermissions_entityId :: Lens' PutMailboxPermissions Text

The identifier of the user, group, or resource for which to update mailbox permissions.

putMailboxPermissions_granteeId :: Lens' PutMailboxPermissions Text

The identifier of the user, group, or resource to which to grant the permissions.

putMailboxPermissions_permissionValues :: Lens' PutMailboxPermissions [PermissionType]

The permissions granted to the grantee. SEND_AS allows the grantee to send email as the owner of the mailbox (the grantee is not mentioned on these emails). SEND_ON_BEHALF allows the grantee to send email on behalf of the owner of the mailbox (the grantee is not mentioned as the physical sender of these emails). FULL_ACCESS allows the grantee full access to the mailbox, irrespective of other folder-level permissions set on the mailbox.

putMailboxPermissionsResponse_httpStatus :: Lens' PutMailboxPermissionsResponse Int

The response's http status code.

GetMobileDeviceAccessEffect

getMobileDeviceAccessEffect_deviceModel :: Lens' GetMobileDeviceAccessEffect (Maybe Text)

Device model the simulated user will report.

getMobileDeviceAccessEffect_deviceUserAgent :: Lens' GetMobileDeviceAccessEffect (Maybe Text)

Device user agent the simulated user will report.

getMobileDeviceAccessEffect_deviceOperatingSystem :: Lens' GetMobileDeviceAccessEffect (Maybe Text)

Device operating system the simulated user will report.

getMobileDeviceAccessEffect_deviceType :: Lens' GetMobileDeviceAccessEffect (Maybe Text)

Device type the simulated user will report.

getMobileDeviceAccessEffect_organizationId :: Lens' GetMobileDeviceAccessEffect Text

The Amazon WorkMail organization to simulate the access effect for.

getMobileDeviceAccessEffectResponse_matchedRules :: Lens' GetMobileDeviceAccessEffectResponse (Maybe [MobileDeviceAccessMatchedRule])

A list of the rules which matched the simulated user input and produced the effect.

getMobileDeviceAccessEffectResponse_effect :: Lens' GetMobileDeviceAccessEffectResponse (Maybe MobileDeviceAccessRuleEffect)

The effect of the simulated access, ALLOW or DENY, after evaluating mobile device access rules in the Amazon WorkMail organization for the simulated user parameters.

getMobileDeviceAccessEffectResponse_httpStatus :: Lens' GetMobileDeviceAccessEffectResponse Int

The response's http status code.

RegisterToWorkMail

registerToWorkMail_organizationId :: Lens' RegisterToWorkMail Text

The identifier for the organization under which the user, group, or resource exists.

registerToWorkMail_entityId :: Lens' RegisterToWorkMail Text

The identifier for the user, group, or resource to be updated.

registerToWorkMail_email :: Lens' RegisterToWorkMail Text

The email for the user, group, or resource to be updated.

registerToWorkMailResponse_httpStatus :: Lens' RegisterToWorkMailResponse Int

The response's http status code.

DeleteOrganization

deleteOrganization_clientToken :: Lens' DeleteOrganization (Maybe Text)

The idempotency token associated with the request.

deleteOrganization_organizationId :: Lens' DeleteOrganization Text

The organization ID.

deleteOrganization_deleteDirectory :: Lens' DeleteOrganization Bool

If true, deletes the AWS Directory Service directory associated with the organization.

deleteOrganizationResponse_organizationId :: Lens' DeleteOrganizationResponse (Maybe Text)

The organization ID.

deleteOrganizationResponse_state :: Lens' DeleteOrganizationResponse (Maybe Text)

The state of the organization.

deleteOrganizationResponse_httpStatus :: Lens' DeleteOrganizationResponse Int

The response's http status code.

DescribeMailboxExportJob

describeMailboxExportJob_jobId :: Lens' DescribeMailboxExportJob Text

The mailbox export job ID.

describeMailboxExportJob_organizationId :: Lens' DescribeMailboxExportJob Text

The organization ID.

describeMailboxExportJobResponse_estimatedProgress :: Lens' DescribeMailboxExportJobResponse (Maybe Natural)

The estimated progress of the mailbox export job, in percentage points.

describeMailboxExportJobResponse_roleArn :: Lens' DescribeMailboxExportJobResponse (Maybe Text)

The ARN of the AWS Identity and Access Management (IAM) role that grants write permission to the Amazon Simple Storage Service (Amazon S3) bucket.

describeMailboxExportJobResponse_entityId :: Lens' DescribeMailboxExportJobResponse (Maybe Text)

The identifier of the user or resource associated with the mailbox.

describeMailboxExportJobResponse_s3Path :: Lens' DescribeMailboxExportJobResponse (Maybe Text)

The path to the S3 bucket and file that the mailbox export job is exporting to.

describeMailboxExportJobResponse_startTime :: Lens' DescribeMailboxExportJobResponse (Maybe UTCTime)

The mailbox export job start timestamp.

describeMailboxExportJobResponse_endTime :: Lens' DescribeMailboxExportJobResponse (Maybe UTCTime)

The mailbox export job end timestamp.

describeMailboxExportJobResponse_kmsKeyArn :: Lens' DescribeMailboxExportJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the symmetric AWS Key Management Service (AWS KMS) key that encrypts the exported mailbox content.

describeMailboxExportJobResponse_state :: Lens' DescribeMailboxExportJobResponse (Maybe MailboxExportJobState)

The state of the mailbox export job.

describeMailboxExportJobResponse_s3BucketName :: Lens' DescribeMailboxExportJobResponse (Maybe Text)

The name of the S3 bucket.

describeMailboxExportJobResponse_description :: Lens' DescribeMailboxExportJobResponse (Maybe Text)

The mailbox export job description.

describeMailboxExportJobResponse_errorInfo :: Lens' DescribeMailboxExportJobResponse (Maybe Text)

Error information for failed mailbox export jobs.

describeMailboxExportJobResponse_s3Prefix :: Lens' DescribeMailboxExportJobResponse (Maybe Text)

The S3 bucket prefix.

describeMailboxExportJobResponse_httpStatus :: Lens' DescribeMailboxExportJobResponse Int

The response's http status code.

AssociateDelegateToResource

associateDelegateToResource_organizationId :: Lens' AssociateDelegateToResource Text

The organization under which the resource exists.

associateDelegateToResource_resourceId :: Lens' AssociateDelegateToResource Text

The resource for which members (users or groups) are associated.

associateDelegateToResource_entityId :: Lens' AssociateDelegateToResource Text

The member (user or group) to associate to the resource.

associateDelegateToResourceResponse_httpStatus :: Lens' AssociateDelegateToResourceResponse Int

The response's http status code.

ListOrganizations

listOrganizations_nextToken :: Lens' ListOrganizations (Maybe Text)

The token to use to retrieve the next page of results. The first call does not contain any tokens.

listOrganizations_maxResults :: Lens' ListOrganizations (Maybe Natural)

The maximum number of results to return in a single call.

listOrganizationsResponse_nextToken :: Lens' ListOrganizationsResponse (Maybe Text)

The token to use to retrieve the next page of results. The value is "null" when there are no more results to return.

listOrganizationsResponse_organizationSummaries :: Lens' ListOrganizationsResponse (Maybe [OrganizationSummary])

The overview of owned organizations presented as a list of organization summaries.

listOrganizationsResponse_httpStatus :: Lens' ListOrganizationsResponse Int

The response's http status code.

UpdatePrimaryEmailAddress

updatePrimaryEmailAddress_organizationId :: Lens' UpdatePrimaryEmailAddress Text

The organization that contains the user, group, or resource to update.

updatePrimaryEmailAddress_entityId :: Lens' UpdatePrimaryEmailAddress Text

The user, group, or resource to update.

updatePrimaryEmailAddress_email :: Lens' UpdatePrimaryEmailAddress Text

The value of the email to be updated as primary.

updatePrimaryEmailAddressResponse_httpStatus :: Lens' UpdatePrimaryEmailAddressResponse Int

The response's http status code.

ListAccessControlRules

listAccessControlRules_organizationId :: Lens' ListAccessControlRules Text

The identifier for the organization.

listAccessControlRulesResponse_rules :: Lens' ListAccessControlRulesResponse (Maybe [AccessControlRule])

The access control rules.

listAccessControlRulesResponse_httpStatus :: Lens' ListAccessControlRulesResponse Int

The response's http status code.

PutAccessControlRule

putAccessControlRule_notIpRanges :: Lens' PutAccessControlRule (Maybe [Text])

IPv4 CIDR ranges to exclude from the rule.

putAccessControlRule_actions :: Lens' PutAccessControlRule (Maybe [Text])

Access protocol actions to include in the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

putAccessControlRule_ipRanges :: Lens' PutAccessControlRule (Maybe [Text])

IPv4 CIDR ranges to include in the rule.

putAccessControlRule_userIds :: Lens' PutAccessControlRule (Maybe [Text])

User IDs to include in the rule.

putAccessControlRule_notActions :: Lens' PutAccessControlRule (Maybe [Text])

Access protocol actions to exclude from the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

putAccessControlRule_notUserIds :: Lens' PutAccessControlRule (Maybe [Text])

User IDs to exclude from the rule.

putAccessControlRule_name :: Lens' PutAccessControlRule Text

The rule name.

putAccessControlRule_effect :: Lens' PutAccessControlRule AccessControlRuleEffect

The rule effect.

putAccessControlRule_description :: Lens' PutAccessControlRule Text

The rule description.

putAccessControlRule_organizationId :: Lens' PutAccessControlRule Text

The identifier of the organization.

putAccessControlRuleResponse_httpStatus :: Lens' PutAccessControlRuleResponse Int

The response's http status code.

DescribeUser

describeUser_organizationId :: Lens' DescribeUser Text

The identifier for the organization under which the user exists.

describeUser_userId :: Lens' DescribeUser Text

The identifier for the user to be described.

describeUserResponse_enabledDate :: Lens' DescribeUserResponse (Maybe UTCTime)

The date and time at which the user was enabled for Amazon WorkMail usage, in UNIX epoch time format.

describeUserResponse_userRole :: Lens' DescribeUserResponse (Maybe UserRole)

In certain cases, other entities are modeled as users. If interoperability is enabled, resources are imported into Amazon WorkMail as users. Because different WorkMail organizations rely on different directory types, administrators can distinguish between an unregistered user (account is disabled and has a user role) and the directory administrators. The values are USER, RESOURCE, and SYSTEM_USER.

describeUserResponse_name :: Lens' DescribeUserResponse (Maybe Text)

The name for the user.

describeUserResponse_state :: Lens' DescribeUserResponse (Maybe EntityState)

The state of a user: enabled (registered to Amazon WorkMail) or disabled (deregistered or never registered to WorkMail).

describeUserResponse_disabledDate :: Lens' DescribeUserResponse (Maybe UTCTime)

The date and time at which the user was disabled for Amazon WorkMail usage, in UNIX epoch time format.

describeUserResponse_email :: Lens' DescribeUserResponse (Maybe Text)

The email of the user.

describeUserResponse_userId :: Lens' DescribeUserResponse (Maybe Text)

The identifier for the described user.

describeUserResponse_displayName :: Lens' DescribeUserResponse (Maybe Text)

The display name of the user.

describeUserResponse_httpStatus :: Lens' DescribeUserResponse Int

The response's http status code.

CancelMailboxExportJob

cancelMailboxExportJob_clientToken :: Lens' CancelMailboxExportJob Text

The idempotency token for the client request.

cancelMailboxExportJob_jobId :: Lens' CancelMailboxExportJob Text

The job ID.

cancelMailboxExportJob_organizationId :: Lens' CancelMailboxExportJob Text

The organization ID.

cancelMailboxExportJobResponse_httpStatus :: Lens' CancelMailboxExportJobResponse Int

The response's http status code.

CreateMobileDeviceAccessRule

createMobileDeviceAccessRule_notDeviceModels :: Lens' CreateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device models that the rule will not match. All other device models will match.

createMobileDeviceAccessRule_notDeviceOperatingSystems :: Lens' CreateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device operating systems that the rule will not match. All other device operating systems will match.

createMobileDeviceAccessRule_deviceUserAgents :: Lens' CreateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device user agents that the rule will match.

createMobileDeviceAccessRule_deviceModels :: Lens' CreateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device models that the rule will match.

createMobileDeviceAccessRule_notDeviceTypes :: Lens' CreateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device types that the rule will not match. All other device types will match.

createMobileDeviceAccessRule_description :: Lens' CreateMobileDeviceAccessRule (Maybe Text)

The rule description.

createMobileDeviceAccessRule_notDeviceUserAgents :: Lens' CreateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device user agents that the rule will not match. All other device user agents will match.

createMobileDeviceAccessRule_deviceTypes :: Lens' CreateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device types that the rule will match.

createMobileDeviceAccessRule_deviceOperatingSystems :: Lens' CreateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device operating systems that the rule will match.

createMobileDeviceAccessRule_clientToken :: Lens' CreateMobileDeviceAccessRule (Maybe Text)

The idempotency token for the client request.

createMobileDeviceAccessRule_organizationId :: Lens' CreateMobileDeviceAccessRule Text

The Amazon WorkMail organization under which the rule will be created.

createMobileDeviceAccessRule_name :: Lens' CreateMobileDeviceAccessRule Text

The rule name.

createMobileDeviceAccessRule_effect :: Lens' CreateMobileDeviceAccessRule MobileDeviceAccessRuleEffect

The effect of the rule when it matches. Allowed values are ALLOW or DENY.

createMobileDeviceAccessRuleResponse_mobileDeviceAccessRuleId :: Lens' CreateMobileDeviceAccessRuleResponse (Maybe Text)

The identifier for the newly created mobile device access rule.

createMobileDeviceAccessRuleResponse_httpStatus :: Lens' CreateMobileDeviceAccessRuleResponse Int

The response's http status code.

DeleteGroup

deleteGroup_organizationId :: Lens' DeleteGroup Text

The organization that contains the group.

deleteGroup_groupId :: Lens' DeleteGroup Text

The identifier of the group to be deleted.

deleteGroupResponse_httpStatus :: Lens' DeleteGroupResponse Int

The response's http status code.

ListGroupMembers

listGroupMembers_nextToken :: Lens' ListGroupMembers (Maybe Text)

The token to use to retrieve the next page of results. The first call does not contain any tokens.

listGroupMembers_maxResults :: Lens' ListGroupMembers (Maybe Natural)

The maximum number of results to return in a single call.

listGroupMembers_organizationId :: Lens' ListGroupMembers Text

The identifier for the organization under which the group exists.

listGroupMembers_groupId :: Lens' ListGroupMembers Text

The identifier for the group to which the members (users or groups) are associated.

listGroupMembersResponse_nextToken :: Lens' ListGroupMembersResponse (Maybe Text)

The token to use to retrieve the next page of results. The first call does not contain any tokens.

listGroupMembersResponse_members :: Lens' ListGroupMembersResponse (Maybe [Member])

The members associated to the group.

listGroupMembersResponse_httpStatus :: Lens' ListGroupMembersResponse Int

The response's http status code.

ListMailboxPermissions

listMailboxPermissions_nextToken :: Lens' ListMailboxPermissions (Maybe Text)

The token to use to retrieve the next page of results. The first call does not contain any tokens.

listMailboxPermissions_maxResults :: Lens' ListMailboxPermissions (Maybe Natural)

The maximum number of results to return in a single call.

listMailboxPermissions_organizationId :: Lens' ListMailboxPermissions Text

The identifier of the organization under which the user, group, or resource exists.

listMailboxPermissions_entityId :: Lens' ListMailboxPermissions Text

The identifier of the user, group, or resource for which to list mailbox permissions.

listMailboxPermissionsResponse_nextToken :: Lens' ListMailboxPermissionsResponse (Maybe Text)

The token to use to retrieve the next page of results. The value is "null" when there are no more results to return.

listMailboxPermissionsResponse_permissions :: Lens' ListMailboxPermissionsResponse (Maybe [Permission])

One page of the user, group, or resource mailbox permissions.

listMailboxPermissionsResponse_httpStatus :: Lens' ListMailboxPermissionsResponse Int

The response's http status code.

DeregisterFromWorkMail

deregisterFromWorkMail_organizationId :: Lens' DeregisterFromWorkMail Text

The identifier for the organization under which the Amazon WorkMail entity exists.

deregisterFromWorkMail_entityId :: Lens' DeregisterFromWorkMail Text

The identifier for the member (user or group) to be updated.

deregisterFromWorkMailResponse_httpStatus :: Lens' DeregisterFromWorkMailResponse Int

The response's http status code.

UpdateMailboxQuota

updateMailboxQuota_organizationId :: Lens' UpdateMailboxQuota Text

The identifier for the organization that contains the user for whom to update the mailbox quota.

updateMailboxQuota_userId :: Lens' UpdateMailboxQuota Text

The identifer for the user for whom to update the mailbox quota.

updateMailboxQuota_mailboxQuota :: Lens' UpdateMailboxQuota Natural

The updated mailbox quota, in MB, for the specified user.

updateMailboxQuotaResponse_httpStatus :: Lens' UpdateMailboxQuotaResponse Int

The response's http status code.

UpdateMobileDeviceAccessRule

updateMobileDeviceAccessRule_notDeviceModels :: Lens' UpdateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device models that the updated rule will not match. All other device models will match.

updateMobileDeviceAccessRule_notDeviceOperatingSystems :: Lens' UpdateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device operating systems that the updated rule will not match. All other device operating systems will match.

updateMobileDeviceAccessRule_deviceUserAgents :: Lens' UpdateMobileDeviceAccessRule (Maybe (NonEmpty Text))

User agents that the updated rule will match.

updateMobileDeviceAccessRule_deviceModels :: Lens' UpdateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device models that the updated rule will match.

updateMobileDeviceAccessRule_notDeviceTypes :: Lens' UpdateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device types that the updated rule will not match. All other device types will match.

updateMobileDeviceAccessRule_description :: Lens' UpdateMobileDeviceAccessRule (Maybe Text)

The updated rule description.

updateMobileDeviceAccessRule_notDeviceUserAgents :: Lens' UpdateMobileDeviceAccessRule (Maybe (NonEmpty Text))

User agents that the updated rule will not match. All other user agents will match.

updateMobileDeviceAccessRule_deviceTypes :: Lens' UpdateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device types that the updated rule will match.

updateMobileDeviceAccessRule_deviceOperatingSystems :: Lens' UpdateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device operating systems that the updated rule will match.

updateMobileDeviceAccessRule_organizationId :: Lens' UpdateMobileDeviceAccessRule Text

The Amazon WorkMail organization under which the rule will be updated.

updateMobileDeviceAccessRule_mobileDeviceAccessRuleId :: Lens' UpdateMobileDeviceAccessRule Text

The identifier of the rule to be updated.

updateMobileDeviceAccessRule_name :: Lens' UpdateMobileDeviceAccessRule Text

The updated rule name.

updateMobileDeviceAccessRule_effect :: Lens' UpdateMobileDeviceAccessRule MobileDeviceAccessRuleEffect

The effect of the rule when it matches. Allowed values are ALLOW or DENY.

updateMobileDeviceAccessRuleResponse_httpStatus :: Lens' UpdateMobileDeviceAccessRuleResponse Int

The response's http status code.

DeleteMobileDeviceAccessRule

deleteMobileDeviceAccessRule_organizationId :: Lens' DeleteMobileDeviceAccessRule Text

The Amazon WorkMail organization under which the rule will be deleted.

deleteMobileDeviceAccessRule_mobileDeviceAccessRuleId :: Lens' DeleteMobileDeviceAccessRule Text

The identifier of the rule to be deleted.

deleteMobileDeviceAccessRuleResponse_httpStatus :: Lens' DeleteMobileDeviceAccessRuleResponse Int

The response's http status code.

DisassociateMemberFromGroup

disassociateMemberFromGroup_organizationId :: Lens' DisassociateMemberFromGroup Text

The identifier for the organization under which the group exists.

disassociateMemberFromGroup_groupId :: Lens' DisassociateMemberFromGroup Text

The identifier for the group from which members are removed.

disassociateMemberFromGroup_memberId :: Lens' DisassociateMemberFromGroup Text

The identifier for the member to be removed to the group.

disassociateMemberFromGroupResponse_httpStatus :: Lens' DisassociateMemberFromGroupResponse Int

The response's http status code.

ListResources

listResources_nextToken :: Lens' ListResources (Maybe Text)

The token to use to retrieve the next page of results. The first call does not contain any tokens.

listResources_maxResults :: Lens' ListResources (Maybe Natural)

The maximum number of results to return in a single call.

listResources_organizationId :: Lens' ListResources Text

The identifier for the organization under which the resources exist.

listResourcesResponse_nextToken :: Lens' ListResourcesResponse (Maybe Text)

The token used to paginate through all the organization's resources. While results are still available, it has an associated value. When the last page is reached, the token is empty.

listResourcesResponse_resources :: Lens' ListResourcesResponse (Maybe [Resource])

One page of the organization's resource representation.

listResourcesResponse_httpStatus :: Lens' ListResourcesResponse Int

The response's http status code.

GetMailboxDetails

getMailboxDetails_organizationId :: Lens' GetMailboxDetails Text

The identifier for the organization that contains the user whose mailbox details are being requested.

getMailboxDetails_userId :: Lens' GetMailboxDetails Text

The identifier for the user whose mailbox details are being requested.

getMailboxDetailsResponse_mailboxQuota :: Lens' GetMailboxDetailsResponse (Maybe Natural)

The maximum allowed mailbox size, in MB, for the specified user.

getMailboxDetailsResponse_mailboxSize :: Lens' GetMailboxDetailsResponse (Maybe Double)

The current mailbox size, in MB, for the specified user.

getMailboxDetailsResponse_httpStatus :: Lens' GetMailboxDetailsResponse Int

The response's http status code.

ListUsers

listUsers_nextToken :: Lens' ListUsers (Maybe Text)

The token to use to retrieve the next page of results. The first call does not contain any tokens.

listUsers_maxResults :: Lens' ListUsers (Maybe Natural)

The maximum number of results to return in a single call.

listUsers_organizationId :: Lens' ListUsers Text

The identifier for the organization under which the users exist.

listUsersResponse_nextToken :: Lens' ListUsersResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is `null` when there are no more results to return.

listUsersResponse_users :: Lens' ListUsersResponse (Maybe [User])

The overview of users for an organization.

listUsersResponse_httpStatus :: Lens' ListUsersResponse Int

The response's http status code.

DeleteUser

deleteUser_organizationId :: Lens' DeleteUser Text

The organization that contains the user to be deleted.

deleteUser_userId :: Lens' DeleteUser Text

The identifier of the user to be deleted.

deleteUserResponse_httpStatus :: Lens' DeleteUserResponse Int

The response's http status code.

DeleteRetentionPolicy

deleteRetentionPolicy_organizationId :: Lens' DeleteRetentionPolicy Text

The organization ID.

deleteRetentionPolicy_id :: Lens' DeleteRetentionPolicy Text

The retention policy ID.

deleteRetentionPolicyResponse_httpStatus :: Lens' DeleteRetentionPolicyResponse Int

The response's http status code.

DeleteMailboxPermissions

deleteMailboxPermissions_organizationId :: Lens' DeleteMailboxPermissions Text

The identifier of the organization under which the member (user or group) exists.

deleteMailboxPermissions_entityId :: Lens' DeleteMailboxPermissions Text

The identifier of the member (user or group) that owns the mailbox.

deleteMailboxPermissions_granteeId :: Lens' DeleteMailboxPermissions Text

The identifier of the member (user or group) for which to delete granted permissions.

deleteMailboxPermissionsResponse_httpStatus :: Lens' DeleteMailboxPermissionsResponse Int

The response's http status code.

ResetPassword

resetPassword_organizationId :: Lens' ResetPassword Text

The identifier of the organization that contains the user for which the password is reset.

resetPassword_userId :: Lens' ResetPassword Text

The identifier of the user for whom the password is reset.

resetPassword_password :: Lens' ResetPassword Text

The new password for the user.

resetPasswordResponse_httpStatus :: Lens' ResetPasswordResponse Int

The response's http status code.

ListAliases

listAliases_nextToken :: Lens' ListAliases (Maybe Text)

The token to use to retrieve the next page of results. The first call does not contain any tokens.

listAliases_maxResults :: Lens' ListAliases (Maybe Natural)

The maximum number of results to return in a single call.

listAliases_organizationId :: Lens' ListAliases Text

The identifier for the organization under which the entity exists.

listAliases_entityId :: Lens' ListAliases Text

The identifier for the entity for which to list the aliases.

listAliasesResponse_nextToken :: Lens' ListAliasesResponse (Maybe Text)

The token to use to retrieve the next page of results. The value is "null" when there are no more results to return.

listAliasesResponse_aliases :: Lens' ListAliasesResponse (Maybe [Text])

The entity's paginated aliases.

listAliasesResponse_httpStatus :: Lens' ListAliasesResponse Int

The response's http status code.

DescribeGroup

describeGroup_organizationId :: Lens' DescribeGroup Text

The identifier for the organization under which the group exists.

describeGroup_groupId :: Lens' DescribeGroup Text

The identifier for the group to be described.

describeGroupResponse_enabledDate :: Lens' DescribeGroupResponse (Maybe UTCTime)

The date and time when a user was registered to WorkMail, in UNIX epoch time format.

describeGroupResponse_groupId :: Lens' DescribeGroupResponse (Maybe Text)

The identifier of the described group.

describeGroupResponse_name :: Lens' DescribeGroupResponse (Maybe Text)

The name of the described group.

describeGroupResponse_state :: Lens' DescribeGroupResponse (Maybe EntityState)

The state of the user: enabled (registered to Amazon WorkMail) or disabled (deregistered or never registered to WorkMail).

describeGroupResponse_disabledDate :: Lens' DescribeGroupResponse (Maybe UTCTime)

The date and time when a user was deregistered from WorkMail, in UNIX epoch time format.

describeGroupResponse_email :: Lens' DescribeGroupResponse (Maybe Text)

The email of the described group.

describeGroupResponse_httpStatus :: Lens' DescribeGroupResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceARN :: Lens' ListTagsForResource Text

The resource ARN.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

A list of tag key-value pairs.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

GetAccessControlEffect

getAccessControlEffect_organizationId :: Lens' GetAccessControlEffect Text

The identifier for the organization.

getAccessControlEffect_ipAddress :: Lens' GetAccessControlEffect Text

The IPv4 address.

getAccessControlEffect_action :: Lens' GetAccessControlEffect Text

The access protocol action. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

getAccessControlEffect_userId :: Lens' GetAccessControlEffect Text

The user ID.

getAccessControlEffectResponse_matchedRules :: Lens' GetAccessControlEffectResponse (Maybe [Text])

The rules that match the given parameters, resulting in an effect.

getAccessControlEffectResponse_effect :: Lens' GetAccessControlEffectResponse (Maybe AccessControlRuleEffect)

The rule effect.

getAccessControlEffectResponse_httpStatus :: Lens' GetAccessControlEffectResponse Int

The response's http status code.

CreateAlias

createAlias_organizationId :: Lens' CreateAlias Text

The organization under which the member (user or group) exists.

createAlias_entityId :: Lens' CreateAlias Text

The member (user or group) to which this alias is added.

createAlias_alias :: Lens' CreateAlias Text

The alias to add to the member set.

createAliasResponse_httpStatus :: Lens' CreateAliasResponse Int

The response's http status code.

Types

AccessControlRule

accessControlRule_effect :: Lens' AccessControlRule (Maybe AccessControlRuleEffect)

The rule effect.

accessControlRule_dateCreated :: Lens' AccessControlRule (Maybe UTCTime)

The date that the rule was created.

accessControlRule_notIpRanges :: Lens' AccessControlRule (Maybe [Text])

IPv4 CIDR ranges to exclude from the rule.

accessControlRule_actions :: Lens' AccessControlRule (Maybe [Text])

Access protocol actions to include in the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

accessControlRule_ipRanges :: Lens' AccessControlRule (Maybe [Text])

IPv4 CIDR ranges to include in the rule.

accessControlRule_dateModified :: Lens' AccessControlRule (Maybe UTCTime)

The date that the rule was modified.

accessControlRule_name :: Lens' AccessControlRule (Maybe Text)

The rule name.

accessControlRule_userIds :: Lens' AccessControlRule (Maybe [Text])

User IDs to include in the rule.

accessControlRule_description :: Lens' AccessControlRule (Maybe Text)

The rule description.

accessControlRule_notActions :: Lens' AccessControlRule (Maybe [Text])

Access protocol actions to exclude from the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

accessControlRule_notUserIds :: Lens' AccessControlRule (Maybe [Text])

User IDs to exclude from the rule.

BookingOptions

bookingOptions_autoDeclineConflictingRequests :: Lens' BookingOptions (Maybe Bool)

The resource's ability to automatically decline any conflicting requests.

bookingOptions_autoDeclineRecurringRequests :: Lens' BookingOptions (Maybe Bool)

The resource's ability to automatically decline any recurring requests.

bookingOptions_autoAcceptRequests :: Lens' BookingOptions (Maybe Bool)

The resource's ability to automatically reply to requests. If disabled, delegates must be associated to the resource.

Delegate

delegate_id :: Lens' Delegate Text

The identifier for the user or group associated as the resource's delegate.

delegate_type :: Lens' Delegate MemberType

The type of the delegate: user or group.

Domain

domain_hostedZoneId :: Lens' Domain (Maybe Text)

The hosted zone ID for a domain hosted in Route 53. Required when configuring a domain hosted in Route 53.

domain_domainName :: Lens' Domain (Maybe Text)

The fully qualified domain name.

FolderConfiguration

folderConfiguration_period :: Lens' FolderConfiguration (Maybe Natural)

The period of time at which the folder configuration action is applied.

folderConfiguration_name :: Lens' FolderConfiguration FolderName

The folder name.

folderConfiguration_action :: Lens' FolderConfiguration RetentionAction

The action to take on the folder contents at the end of the folder configuration period.

Group

group_enabledDate :: Lens' Group (Maybe UTCTime)

The date indicating when the group was enabled for Amazon WorkMail use.

group_id :: Lens' Group (Maybe Text)

The identifier of the group.

group_name :: Lens' Group (Maybe Text)

The name of the group.

group_state :: Lens' Group (Maybe EntityState)

The state of the group, which can be ENABLED, DISABLED, or DELETED.

group_disabledDate :: Lens' Group (Maybe UTCTime)

The date indicating when the group was disabled from Amazon WorkMail use.

group_email :: Lens' Group (Maybe Text)

The email of the group.

MailboxExportJob

mailboxExportJob_estimatedProgress :: Lens' MailboxExportJob (Maybe Natural)

The estimated progress of the mailbox export job, in percentage points.

mailboxExportJob_entityId :: Lens' MailboxExportJob (Maybe Text)

The identifier of the user or resource associated with the mailbox.

mailboxExportJob_s3Path :: Lens' MailboxExportJob (Maybe Text)

The path to the S3 bucket and file that the mailbox export job exports to.

mailboxExportJob_startTime :: Lens' MailboxExportJob (Maybe UTCTime)

The mailbox export job start timestamp.

mailboxExportJob_endTime :: Lens' MailboxExportJob (Maybe UTCTime)

The mailbox export job end timestamp.

mailboxExportJob_state :: Lens' MailboxExportJob (Maybe MailboxExportJobState)

The state of the mailbox export job.

mailboxExportJob_s3BucketName :: Lens' MailboxExportJob (Maybe Text)

The name of the S3 bucket.

mailboxExportJob_description :: Lens' MailboxExportJob (Maybe Text)

The mailbox export job description.

mailboxExportJob_jobId :: Lens' MailboxExportJob (Maybe Text)

The identifier of the mailbox export job.

Member

member_enabledDate :: Lens' Member (Maybe UTCTime)

The date indicating when the member was enabled for Amazon WorkMail use.

member_id :: Lens' Member (Maybe Text)

The identifier of the member.

member_name :: Lens' Member (Maybe Text)

The name of the member.

member_state :: Lens' Member (Maybe EntityState)

The state of the member, which can be ENABLED, DISABLED, or DELETED.

member_disabledDate :: Lens' Member (Maybe UTCTime)

The date indicating when the member was disabled from Amazon WorkMail use.

member_type :: Lens' Member (Maybe MemberType)

A member can be a user or group.

MobileDeviceAccessMatchedRule

mobileDeviceAccessMatchedRule_name :: Lens' MobileDeviceAccessMatchedRule (Maybe Text)

Name of a rule that a simulated user matches.

mobileDeviceAccessMatchedRule_mobileDeviceAccessRuleId :: Lens' MobileDeviceAccessMatchedRule (Maybe Text)

Identifier of the rule that a simulated user matches.

MobileDeviceAccessRule

mobileDeviceAccessRule_effect :: Lens' MobileDeviceAccessRule (Maybe MobileDeviceAccessRuleEffect)

The effect of the rule when it matches. Allowed values are ALLOW or DENY.

mobileDeviceAccessRule_dateCreated :: Lens' MobileDeviceAccessRule (Maybe UTCTime)

The date and time at which an access rule was created.

mobileDeviceAccessRule_notDeviceModels :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device models that a rule will not match. All other device models will match.

mobileDeviceAccessRule_notDeviceOperatingSystems :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device operating systems that a rule will not match. All other device types will match.

mobileDeviceAccessRule_deviceUserAgents :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device user agents that a rule will match.

mobileDeviceAccessRule_dateModified :: Lens' MobileDeviceAccessRule (Maybe UTCTime)

The date and time at which an access rule was modified.

mobileDeviceAccessRule_name :: Lens' MobileDeviceAccessRule (Maybe Text)

The name of a mobile access rule.

mobileDeviceAccessRule_mobileDeviceAccessRuleId :: Lens' MobileDeviceAccessRule (Maybe Text)

The ID assigned to a mobile access rule.

mobileDeviceAccessRule_deviceModels :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device models that a rule will match.

mobileDeviceAccessRule_notDeviceTypes :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device types that a rule will not match. All other device types will match.

mobileDeviceAccessRule_description :: Lens' MobileDeviceAccessRule (Maybe Text)

The description of a mobile access rule.

mobileDeviceAccessRule_notDeviceUserAgents :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device user agents that a rule will not match. All other device user agents will match.

mobileDeviceAccessRule_deviceTypes :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device types that a rule will match.

mobileDeviceAccessRule_deviceOperatingSystems :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device operating systems that a rule will match.

OrganizationSummary

organizationSummary_alias :: Lens' OrganizationSummary (Maybe Text)

The alias associated with the organization.

organizationSummary_organizationId :: Lens' OrganizationSummary (Maybe Text)

The identifier associated with the organization.

organizationSummary_defaultMailDomain :: Lens' OrganizationSummary (Maybe Text)

The default email domain associated with the organization.

organizationSummary_state :: Lens' OrganizationSummary (Maybe Text)

The state associated with the organization.

organizationSummary_errorMessage :: Lens' OrganizationSummary (Maybe Text)

The error message associated with the organization. It is only present if unexpected behavior has occurred with regards to the organization. It provides insight or solutions regarding unexpected behavior.

Permission

permission_granteeId :: Lens' Permission Text

The identifier of the user, group, or resource to which the permissions are granted.

permission_granteeType :: Lens' Permission MemberType

The type of user, group, or resource referred to in GranteeId.

permission_permissionValues :: Lens' Permission [PermissionType]

The permissions granted to the grantee. SEND_AS allows the grantee to send email as the owner of the mailbox (the grantee is not mentioned on these emails). SEND_ON_BEHALF allows the grantee to send email on behalf of the owner of the mailbox (the grantee is not mentioned as the physical sender of these emails). FULL_ACCESS allows the grantee full access to the mailbox, irrespective of other folder-level permissions set on the mailbox.

Resource

resource_enabledDate :: Lens' Resource (Maybe UTCTime)

The date indicating when the resource was enabled for Amazon WorkMail use.

resource_id :: Lens' Resource (Maybe Text)

The identifier of the resource.

resource_name :: Lens' Resource (Maybe Text)

The name of the resource.

resource_state :: Lens' Resource (Maybe EntityState)

The state of the resource, which can be ENABLED, DISABLED, or DELETED.

resource_disabledDate :: Lens' Resource (Maybe UTCTime)

The date indicating when the resource was disabled from Amazon WorkMail use.

resource_email :: Lens' Resource (Maybe Text)

The email of the resource.

resource_type :: Lens' Resource (Maybe ResourceType)

The type of the resource: equipment or room.

Tag

tag_key :: Lens' Tag Text

The key of the tag.

tag_value :: Lens' Tag Text

The value of the tag.

User

user_enabledDate :: Lens' User (Maybe UTCTime)

The date indicating when the user was enabled for Amazon WorkMail use.

user_id :: Lens' User (Maybe Text)

The identifier of the user.

user_userRole :: Lens' User (Maybe UserRole)

The role of the user.

user_name :: Lens' User (Maybe Text)

The name of the user.

user_state :: Lens' User (Maybe EntityState)

The state of the user, which can be ENABLED, DISABLED, or DELETED.

user_disabledDate :: Lens' User (Maybe UTCTime)

The date indicating when the user was disabled from Amazon WorkMail use.

user_email :: Lens' User (Maybe Text)

The email of the user.

user_displayName :: Lens' User (Maybe Text)

The display name of the user.