Network.AWS.WorkMail.PutAccessControlRule

Description

Adds a new access control rule for the specified organization. The rule allows or denies access to the organization for the specified IPv4 addresses, access protocol actions, and user IDs. Adding a new rule with the same name as an existing rule replaces the older rule.

Creating a Request

data PutAccessControlRule

See: newPutAccessControlRule smart constructor.

Constructors

PutAccessControlRule'

Fields

userIds :: Maybe [Text]

User IDs to include in the rule.

actions :: Maybe [Text]

Access protocol actions to include in the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

notUserIds :: Maybe [Text]

User IDs to exclude from the rule.

ipRanges :: Maybe [Text]

IPv4 CIDR ranges to include in the rule.

notIpRanges :: Maybe [Text]

IPv4 CIDR ranges to exclude from the rule.

notActions :: Maybe [Text]

Access protocol actions to exclude from the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

name :: Text

The rule name.

effect :: AccessControlRuleEffect

The rule effect.

description :: Text

The rule description.

organizationId :: Text

The identifier of the organization.

newPutAccessControlRule

Arguments

:: Text
$sel:name:PutAccessControlRule'

-> AccessControlRuleEffect
$sel:effect:PutAccessControlRule'

-> Text
$sel:description:PutAccessControlRule'

-> Text
$sel:organizationId:PutAccessControlRule'

-> PutAccessControlRule

Create a value of PutAccessControlRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userIds:PutAccessControlRule', putAccessControlRule_userIds - User IDs to include in the rule.

$sel:actions:PutAccessControlRule', putAccessControlRule_actions - Access protocol actions to include in the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

$sel:notUserIds:PutAccessControlRule', putAccessControlRule_notUserIds - User IDs to exclude from the rule.

$sel:ipRanges:PutAccessControlRule', putAccessControlRule_ipRanges - IPv4 CIDR ranges to include in the rule.

$sel:notIpRanges:PutAccessControlRule', putAccessControlRule_notIpRanges - IPv4 CIDR ranges to exclude from the rule.

$sel:notActions:PutAccessControlRule', putAccessControlRule_notActions - Access protocol actions to exclude from the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

$sel:name:PutAccessControlRule', putAccessControlRule_name - The rule name.

$sel:effect:PutAccessControlRule', putAccessControlRule_effect - The rule effect.

$sel:description:PutAccessControlRule', putAccessControlRule_description - The rule description.

$sel:organizationId:PutAccessControlRule', putAccessControlRule_organizationId - The identifier of the organization.

Request Lenses

putAccessControlRule_userIds :: Lens' PutAccessControlRule (Maybe [Text])

User IDs to include in the rule.

putAccessControlRule_actions :: Lens' PutAccessControlRule (Maybe [Text])

Access protocol actions to include in the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

putAccessControlRule_notUserIds :: Lens' PutAccessControlRule (Maybe [Text])

User IDs to exclude from the rule.

putAccessControlRule_ipRanges :: Lens' PutAccessControlRule (Maybe [Text])

IPv4 CIDR ranges to include in the rule.

putAccessControlRule_notIpRanges :: Lens' PutAccessControlRule (Maybe [Text])

IPv4 CIDR ranges to exclude from the rule.

putAccessControlRule_notActions :: Lens' PutAccessControlRule (Maybe [Text])

Access protocol actions to exclude from the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

putAccessControlRule_name :: Lens' PutAccessControlRule Text

The rule name.

putAccessControlRule_effect :: Lens' PutAccessControlRule AccessControlRuleEffect

The rule effect.

putAccessControlRule_description :: Lens' PutAccessControlRule Text

The rule description.

putAccessControlRule_organizationId :: Lens' PutAccessControlRule Text

The identifier of the organization.

Destructuring the Response

data PutAccessControlRuleResponse

See: newPutAccessControlRuleResponse smart constructor.

Constructors

PutAccessControlRuleResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutAccessControlRuleResponse

Arguments

:: Int
$sel:httpStatus:PutAccessControlRuleResponse'

-> PutAccessControlRuleResponse

Create a value of PutAccessControlRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutAccessControlRuleResponse', putAccessControlRuleResponse_httpStatus - The response's http status code.

Response Lenses

putAccessControlRuleResponse_httpStatus :: Lens' PutAccessControlRuleResponse Int

The response's http status code.