Network.AWS.WorkMail.PutRetentionPolicy

Description

Puts a retention policy to the specified organization.

Creating a Request

data PutRetentionPolicy

See: newPutRetentionPolicy smart constructor.

Constructors

PutRetentionPolicy'

Fields

id :: Maybe Text

The retention policy ID.

description :: Maybe (Sensitive Text)

The retention policy description.

organizationId :: Text

The organization ID.

name :: Text

The retention policy name.

folderConfigurations :: [FolderConfiguration]

The retention policy folder configurations.

newPutRetentionPolicy

Arguments

:: Text
$sel:organizationId:PutRetentionPolicy'

-> Text
$sel:name:PutRetentionPolicy'

-> PutRetentionPolicy

Create a value of PutRetentionPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:PutRetentionPolicy', putRetentionPolicy_id - The retention policy ID.

$sel:description:PutRetentionPolicy', putRetentionPolicy_description - The retention policy description.

$sel:organizationId:PutRetentionPolicy', putRetentionPolicy_organizationId - The organization ID.

$sel:name:PutRetentionPolicy', putRetentionPolicy_name - The retention policy name.

$sel:folderConfigurations:PutRetentionPolicy', putRetentionPolicy_folderConfigurations - The retention policy folder configurations.

Request Lenses

putRetentionPolicy_id :: Lens' PutRetentionPolicy (Maybe Text)

The retention policy ID.

putRetentionPolicy_description :: Lens' PutRetentionPolicy (Maybe Text)

The retention policy description.

putRetentionPolicy_organizationId :: Lens' PutRetentionPolicy Text

The organization ID.

putRetentionPolicy_name :: Lens' PutRetentionPolicy Text

The retention policy name.

putRetentionPolicy_folderConfigurations :: Lens' PutRetentionPolicy [FolderConfiguration]

The retention policy folder configurations.

Destructuring the Response

data PutRetentionPolicyResponse

See: newPutRetentionPolicyResponse smart constructor.

Constructors

PutRetentionPolicyResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutRetentionPolicyResponse

Arguments

:: Int
$sel:httpStatus:PutRetentionPolicyResponse'

-> PutRetentionPolicyResponse

Create a value of PutRetentionPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutRetentionPolicyResponse', putRetentionPolicyResponse_httpStatus - The response's http status code.

Response Lenses

putRetentionPolicyResponse_httpStatus :: Lens' PutRetentionPolicyResponse Int

The response's http status code.