Network.AWS.WorkMail.StartMailboxExportJob

Description

Starts a mailbox export job to export MIME-format email messages and calendar items from the specified mailbox to the specified Amazon Simple Storage Service (Amazon S3) bucket. For more information, see Exporting mailbox content in the Amazon WorkMail Administrator Guide.

Creating a Request

data StartMailboxExportJob

See: newStartMailboxExportJob smart constructor.

Constructors

StartMailboxExportJob'

Fields

description :: Maybe Text

The mailbox export job description.

clientToken :: Text

The idempotency token for the client request.

organizationId :: Text

The identifier associated with the organization.

entityId :: Text

The identifier of the user or resource associated with the mailbox.

roleArn :: Text

The ARN of the AWS Identity and Access Management (IAM) role that grants write permission to the S3 bucket.

kmsKeyArn :: Text

The Amazon Resource Name (ARN) of the symmetric AWS Key Management Service (AWS KMS) key that encrypts the exported mailbox content.

s3BucketName :: Text

The name of the S3 bucket.

s3Prefix :: Text

The S3 bucket prefix.

newStartMailboxExportJob

Arguments

:: Text
$sel:clientToken:StartMailboxExportJob'

-> Text
$sel:organizationId:StartMailboxExportJob'

-> Text
$sel:entityId:StartMailboxExportJob'

-> Text
$sel:roleArn:StartMailboxExportJob'

-> Text
$sel:kmsKeyArn:StartMailboxExportJob'

-> Text
$sel:s3BucketName:StartMailboxExportJob'

-> Text
$sel:s3Prefix:StartMailboxExportJob'

-> StartMailboxExportJob

Create a value of StartMailboxExportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:StartMailboxExportJob', startMailboxExportJob_description - The mailbox export job description.

$sel:clientToken:StartMailboxExportJob', startMailboxExportJob_clientToken - The idempotency token for the client request.

$sel:organizationId:StartMailboxExportJob', startMailboxExportJob_organizationId - The identifier associated with the organization.

$sel:entityId:StartMailboxExportJob', startMailboxExportJob_entityId - The identifier of the user or resource associated with the mailbox.

$sel:roleArn:StartMailboxExportJob', startMailboxExportJob_roleArn - The ARN of the AWS Identity and Access Management (IAM) role that grants write permission to the S3 bucket.

$sel:kmsKeyArn:StartMailboxExportJob', startMailboxExportJob_kmsKeyArn - The Amazon Resource Name (ARN) of the symmetric AWS Key Management Service (AWS KMS) key that encrypts the exported mailbox content.

$sel:s3BucketName:StartMailboxExportJob', startMailboxExportJob_s3BucketName - The name of the S3 bucket.

$sel:s3Prefix:StartMailboxExportJob', startMailboxExportJob_s3Prefix - The S3 bucket prefix.

Request Lenses

startMailboxExportJob_description :: Lens' StartMailboxExportJob (Maybe Text)

The mailbox export job description.

startMailboxExportJob_clientToken :: Lens' StartMailboxExportJob Text

The idempotency token for the client request.

startMailboxExportJob_organizationId :: Lens' StartMailboxExportJob Text

The identifier associated with the organization.

startMailboxExportJob_entityId :: Lens' StartMailboxExportJob Text

The identifier of the user or resource associated with the mailbox.

startMailboxExportJob_roleArn :: Lens' StartMailboxExportJob Text

The ARN of the AWS Identity and Access Management (IAM) role that grants write permission to the S3 bucket.

startMailboxExportJob_kmsKeyArn :: Lens' StartMailboxExportJob Text

The Amazon Resource Name (ARN) of the symmetric AWS Key Management Service (AWS KMS) key that encrypts the exported mailbox content.

startMailboxExportJob_s3BucketName :: Lens' StartMailboxExportJob Text

The name of the S3 bucket.

startMailboxExportJob_s3Prefix :: Lens' StartMailboxExportJob Text

The S3 bucket prefix.

Destructuring the Response

data StartMailboxExportJobResponse

See: newStartMailboxExportJobResponse smart constructor.

Constructors

StartMailboxExportJobResponse'

Fields

jobId :: Maybe Text

The job ID.

httpStatus :: Int

The response's http status code.

newStartMailboxExportJobResponse

Arguments

:: Int
$sel:httpStatus:StartMailboxExportJobResponse'

-> StartMailboxExportJobResponse

Create a value of StartMailboxExportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:StartMailboxExportJobResponse', startMailboxExportJobResponse_jobId - The job ID.

$sel:httpStatus:StartMailboxExportJobResponse', startMailboxExportJobResponse_httpStatus - The response's http status code.

Response Lenses

startMailboxExportJobResponse_jobId :: Lens' StartMailboxExportJobResponse (Maybe Text)

The job ID.

startMailboxExportJobResponse_httpStatus :: Lens' StartMailboxExportJobResponse Int

The response's http status code.