Network.AWS.WorkMail.Types

Description

Service Configuration

defaultService :: Service

API version 2017-10-01 of the Amazon WorkMail SDK configuration.

Errors

_DirectoryUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The directory is unavailable. It might be located in another Region or deleted.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

One or more of the input parameters don't match the service's restrictions.

_UnsupportedOperationException :: AsError a => Getting (First ServiceError) a ServiceError

You can't perform a write operation against a read-only directory.

_DirectoryServiceAuthenticationFailedException :: AsError a => Getting (First ServiceError) a ServiceError

The directory service doesn't recognize the credentials supplied by WorkMail.

_OrganizationStateException :: AsError a => Getting (First ServiceError) a ServiceError

The organization must have a valid state to perform certain operations on the organization or its members.

_EntityStateException :: AsError a => Getting (First ServiceError) a ServiceError

You are performing an operation on a user, group, or resource that isn't in the expected state, such as trying to delete an active user.

_InvalidConfigurationException :: AsError a => Getting (First ServiceError) a ServiceError

The configuration for a resource isn't valid. A resource must either be able to auto-respond to requests or have at least one delegate associated that can do so on its behalf.

_TooManyTagsException :: AsError a => Getting (First ServiceError) a ServiceError

The resource can have up to 50 user-applied tags.

_MailDomainStateException :: AsError a => Getting (First ServiceError) a ServiceError

After a domain has been added to the organization, it must be verified. The domain is not yet verified.

_ReservedNameException :: AsError a => Getting (First ServiceError) a ServiceError

This user, group, or resource name is not allowed in Amazon WorkMail.

_OrganizationNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

An operation received a valid organization identifier that either doesn't belong or exist in the system.

_EntityNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The identifier supplied for the user, group, or resource does not exist in your organization.

_EntityAlreadyRegisteredException :: AsError a => Getting (First ServiceError) a ServiceError

The user, group, or resource that you're trying to register is already registered.

_DirectoryInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The directory is already in use by another WorkMail organization in the same account and Region.

_MailDomainInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The domain you're trying to change is in use by another user or organization in your account. See the error message for details.

_MailDomainNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The domain specified is not found in your organization.

_InvalidCustomSesConfigurationException :: AsError a => Getting (First ServiceError) a ServiceError

You SES configuration has customizations that Amazon WorkMail cannot save. The error message lists the invalid setting. For examples of invalid settings, refer to CreateReceiptRule.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resource cannot be found.

_EmailAddressInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The email address that you're trying to assign is already created for a different user, group, or resource.

_NameAvailabilityException :: AsError a => Getting (First ServiceError) a ServiceError

The user, group, or resource name isn't unique in Amazon WorkMail.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request exceeds the limit of the resource.

_InvalidPasswordException :: AsError a => Getting (First ServiceError) a ServiceError

The supplied password doesn't match the minimum security constraints, such as length or use of special characters.

AccessControlRuleEffect

newtype AccessControlRuleEffect

Constructors

AccessControlRuleEffect'

Fields

fromAccessControlRuleEffect :: Text

Bundled Patterns

pattern AccessControlRuleEffect_ALLOW :: AccessControlRuleEffect

pattern AccessControlRuleEffect_DENY :: AccessControlRuleEffect

DnsRecordVerificationStatus

newtype DnsRecordVerificationStatus

Constructors

DnsRecordVerificationStatus'

Fields

fromDnsRecordVerificationStatus :: Text

Bundled Patterns

pattern DnsRecordVerificationStatus_FAILED :: DnsRecordVerificationStatus

pattern DnsRecordVerificationStatus_PENDING :: DnsRecordVerificationStatus

pattern DnsRecordVerificationStatus_VERIFIED :: DnsRecordVerificationStatus

EntityState

newtype EntityState

Constructors

EntityState'

Fields

fromEntityState :: Text

Bundled Patterns

pattern EntityState_DELETED :: EntityState

pattern EntityState_DISABLED :: EntityState

pattern EntityState_ENABLED :: EntityState

FolderName

newtype FolderName

Constructors

FolderName'

Fields

fromFolderName :: Text

Bundled Patterns

pattern FolderName_DELETED_ITEMS :: FolderName

pattern FolderName_DRAFTS :: FolderName

pattern FolderName_INBOX :: FolderName

pattern FolderName_JUNK_EMAIL :: FolderName

pattern FolderName_SENT_ITEMS :: FolderName

MailboxExportJobState

newtype MailboxExportJobState

Constructors

MailboxExportJobState'

Fields

fromMailboxExportJobState :: Text

Bundled Patterns

pattern MailboxExportJobState_CANCELLED :: MailboxExportJobState

pattern MailboxExportJobState_COMPLETED :: MailboxExportJobState

pattern MailboxExportJobState_FAILED :: MailboxExportJobState

pattern MailboxExportJobState_RUNNING :: MailboxExportJobState

MemberType

newtype MemberType

Constructors

MemberType'

Fields

fromMemberType :: Text

Bundled Patterns

pattern MemberType_GROUP :: MemberType

pattern MemberType_USER :: MemberType

MobileDeviceAccessRuleEffect

newtype MobileDeviceAccessRuleEffect

Constructors

MobileDeviceAccessRuleEffect'

Fields

fromMobileDeviceAccessRuleEffect :: Text

Bundled Patterns

pattern MobileDeviceAccessRuleEffect_ALLOW :: MobileDeviceAccessRuleEffect

pattern MobileDeviceAccessRuleEffect_DENY :: MobileDeviceAccessRuleEffect

PermissionType

newtype PermissionType

Constructors

PermissionType'

Fields

fromPermissionType :: Text

Bundled Patterns

pattern PermissionType_FULL_ACCESS :: PermissionType

pattern PermissionType_SEND_AS :: PermissionType

pattern PermissionType_SEND_ON_BEHALF :: PermissionType

ResourceType

newtype ResourceType

Constructors

ResourceType'

Fields

fromResourceType :: Text

Bundled Patterns

pattern ResourceType_EQUIPMENT :: ResourceType

pattern ResourceType_ROOM :: ResourceType

RetentionAction

newtype RetentionAction

Constructors

RetentionAction'

Fields

fromRetentionAction :: Text

Bundled Patterns

pattern RetentionAction_DELETE :: RetentionAction

pattern RetentionAction_NONE :: RetentionAction

pattern RetentionAction_PERMANENTLY_DELETE :: RetentionAction

UserRole

newtype UserRole

Constructors

UserRole'

Fields

fromUserRole :: Text

Bundled Patterns

pattern UserRole_RESOURCE :: UserRole

pattern UserRole_SYSTEM_USER :: UserRole

pattern UserRole_USER :: UserRole

AccessControlRule

data AccessControlRule

A rule that controls access to an Amazon WorkMail organization.

See: newAccessControlRule smart constructor.

Constructors

AccessControlRule'

Fields

effect :: Maybe AccessControlRuleEffect

The rule effect.

userIds :: Maybe [Text]

User IDs to include in the rule.

actions :: Maybe [Text]

Access protocol actions to include in the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

dateCreated :: Maybe POSIX

The date that the rule was created.

name :: Maybe Text

The rule name.

notUserIds :: Maybe [Text]

User IDs to exclude from the rule.

dateModified :: Maybe POSIX

The date that the rule was modified.

ipRanges :: Maybe [Text]

IPv4 CIDR ranges to include in the rule.

notIpRanges :: Maybe [Text]

IPv4 CIDR ranges to exclude from the rule.

notActions :: Maybe [Text]

Access protocol actions to exclude from the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

description :: Maybe Text

The rule description.

newAccessControlRule :: AccessControlRule

Create a value of AccessControlRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:effect:AccessControlRule', accessControlRule_effect - The rule effect.

$sel:userIds:AccessControlRule', accessControlRule_userIds - User IDs to include in the rule.

$sel:actions:AccessControlRule', accessControlRule_actions - Access protocol actions to include in the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

$sel:dateCreated:AccessControlRule', accessControlRule_dateCreated - The date that the rule was created.

$sel:name:AccessControlRule', accessControlRule_name - The rule name.

$sel:notUserIds:AccessControlRule', accessControlRule_notUserIds - User IDs to exclude from the rule.

$sel:dateModified:AccessControlRule', accessControlRule_dateModified - The date that the rule was modified.

$sel:ipRanges:AccessControlRule', accessControlRule_ipRanges - IPv4 CIDR ranges to include in the rule.

$sel:notIpRanges:AccessControlRule', accessControlRule_notIpRanges - IPv4 CIDR ranges to exclude from the rule.

$sel:notActions:AccessControlRule', accessControlRule_notActions - Access protocol actions to exclude from the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

$sel:description:AccessControlRule', accessControlRule_description - The rule description.

accessControlRule_effect :: Lens' AccessControlRule (Maybe AccessControlRuleEffect)

The rule effect.

accessControlRule_userIds :: Lens' AccessControlRule (Maybe [Text])

User IDs to include in the rule.

accessControlRule_actions :: Lens' AccessControlRule (Maybe [Text])

Access protocol actions to include in the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

accessControlRule_dateCreated :: Lens' AccessControlRule (Maybe UTCTime)

The date that the rule was created.

accessControlRule_name :: Lens' AccessControlRule (Maybe Text)

The rule name.

accessControlRule_notUserIds :: Lens' AccessControlRule (Maybe [Text])

User IDs to exclude from the rule.

accessControlRule_dateModified :: Lens' AccessControlRule (Maybe UTCTime)

The date that the rule was modified.

accessControlRule_ipRanges :: Lens' AccessControlRule (Maybe [Text])

IPv4 CIDR ranges to include in the rule.

accessControlRule_notIpRanges :: Lens' AccessControlRule (Maybe [Text])

IPv4 CIDR ranges to exclude from the rule.

accessControlRule_notActions :: Lens' AccessControlRule (Maybe [Text])

Access protocol actions to exclude from the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

accessControlRule_description :: Lens' AccessControlRule (Maybe Text)

The rule description.

BookingOptions

data BookingOptions

At least one delegate must be associated to the resource to disable automatic replies from the resource.

See: newBookingOptions smart constructor.

Constructors

BookingOptions'

Fields

autoDeclineConflictingRequests :: Maybe Bool

The resource's ability to automatically decline any conflicting requests.

autoDeclineRecurringRequests :: Maybe Bool

The resource's ability to automatically decline any recurring requests.

autoAcceptRequests :: Maybe Bool

The resource's ability to automatically reply to requests. If disabled, delegates must be associated to the resource.

newBookingOptions :: BookingOptions

Create a value of BookingOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoDeclineConflictingRequests:BookingOptions', bookingOptions_autoDeclineConflictingRequests - The resource's ability to automatically decline any conflicting requests.

$sel:autoDeclineRecurringRequests:BookingOptions', bookingOptions_autoDeclineRecurringRequests - The resource's ability to automatically decline any recurring requests.

$sel:autoAcceptRequests:BookingOptions', bookingOptions_autoAcceptRequests - The resource's ability to automatically reply to requests. If disabled, delegates must be associated to the resource.

bookingOptions_autoDeclineConflictingRequests :: Lens' BookingOptions (Maybe Bool)

The resource's ability to automatically decline any conflicting requests.

bookingOptions_autoDeclineRecurringRequests :: Lens' BookingOptions (Maybe Bool)

The resource's ability to automatically decline any recurring requests.

bookingOptions_autoAcceptRequests :: Lens' BookingOptions (Maybe Bool)

The resource's ability to automatically reply to requests. If disabled, delegates must be associated to the resource.

Delegate

data Delegate

The name of the attribute, which is one of the values defined in the UserAttribute enumeration.

See: newDelegate smart constructor.

Constructors

Delegate'

Fields

id :: Text

The identifier for the user or group associated as the resource's delegate.

type' :: MemberType

The type of the delegate: user or group.

newDelegate

Arguments

:: Text
$sel:id:Delegate'

-> MemberType
$sel:type':Delegate'

-> Delegate

Create a value of Delegate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:Delegate', delegate_id - The identifier for the user or group associated as the resource's delegate.

$sel:type':Delegate', delegate_type - The type of the delegate: user or group.

delegate_id :: Lens' Delegate Text

The identifier for the user or group associated as the resource's delegate.

delegate_type :: Lens' Delegate MemberType

The type of the delegate: user or group.

DnsRecord

data DnsRecord

A DNS record uploaded to your DNS provider.

See: newDnsRecord smart constructor.

Constructors

DnsRecord'

Fields

hostname :: Maybe Text

The DNS hostname.- For example, domain.example.com.

value :: Maybe Text

The value returned by the DNS for a query to that hostname and record type.

type' :: Maybe Text

The RFC 1035 record type. Possible values: CNAME, A, MX.

newDnsRecord :: DnsRecord

Create a value of DnsRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostname:DnsRecord', dnsRecord_hostname - The DNS hostname.- For example, domain.example.com.

$sel:value:DnsRecord', dnsRecord_value - The value returned by the DNS for a query to that hostname and record type.

$sel:type':DnsRecord', dnsRecord_type - The RFC 1035 record type. Possible values: CNAME, A, MX.

dnsRecord_hostname :: Lens' DnsRecord (Maybe Text)

The DNS hostname.- For example, domain.example.com.

dnsRecord_value :: Lens' DnsRecord (Maybe Text)

The value returned by the DNS for a query to that hostname and record type.

dnsRecord_type :: Lens' DnsRecord (Maybe Text)

The RFC 1035 record type. Possible values: CNAME, A, MX.

Domain

data Domain

The domain to associate with an Amazon WorkMail organization.

When you configure a domain hosted in Amazon Route 53 (Route 53), all recommended DNS records are added to the organization when you create it. For more information, see Adding a domain in the Amazon WorkMail Administrator Guide.

See: newDomain smart constructor.

Constructors

Domain'

Fields

hostedZoneId :: Maybe Text

The hosted zone ID for a domain hosted in Route 53. Required when configuring a domain hosted in Route 53.

domainName :: Maybe Text

The fully qualified domain name.

newDomain :: Domain

Create a value of Domain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostedZoneId:Domain', domain_hostedZoneId - The hosted zone ID for a domain hosted in Route 53. Required when configuring a domain hosted in Route 53.

$sel:domainName:Domain', domain_domainName - The fully qualified domain name.

domain_hostedZoneId :: Lens' Domain (Maybe Text)

The hosted zone ID for a domain hosted in Route 53. Required when configuring a domain hosted in Route 53.

domain_domainName :: Lens' Domain (Maybe Text)

The fully qualified domain name.

FolderConfiguration

data FolderConfiguration

The configuration applied to an organization's folders by its retention policy.

See: newFolderConfiguration smart constructor.

Constructors

FolderConfiguration'

Fields

period :: Maybe Natural

The number of days for which the folder-configuration action applies.

name :: FolderName

The folder name.

action :: RetentionAction

The action to take on the folder contents at the end of the folder configuration period.

newFolderConfiguration

Arguments

:: FolderName
$sel:name:FolderConfiguration'

-> RetentionAction
$sel:action:FolderConfiguration'

-> FolderConfiguration

Create a value of FolderConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:period:FolderConfiguration', folderConfiguration_period - The number of days for which the folder-configuration action applies.

$sel:name:FolderConfiguration', folderConfiguration_name - The folder name.

$sel:action:FolderConfiguration', folderConfiguration_action - The action to take on the folder contents at the end of the folder configuration period.

folderConfiguration_period :: Lens' FolderConfiguration (Maybe Natural)

The number of days for which the folder-configuration action applies.

folderConfiguration_name :: Lens' FolderConfiguration FolderName

The folder name.

folderConfiguration_action :: Lens' FolderConfiguration RetentionAction

The action to take on the folder contents at the end of the folder configuration period.

Group

data Group

The representation of an Amazon WorkMail group.

See: newGroup smart constructor.

Constructors

Group'

Fields

email :: Maybe Text

The email of the group.

state :: Maybe EntityState

The state of the group, which can be ENABLED, DISABLED, or DELETED.

disabledDate :: Maybe POSIX

The date indicating when the group was disabled from Amazon WorkMail use.

name :: Maybe Text

The name of the group.

id :: Maybe Text

The identifier of the group.

enabledDate :: Maybe POSIX

The date indicating when the group was enabled for Amazon WorkMail use.

newGroup :: Group

Create a value of Group with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:email:Group', group_email - The email of the group.

$sel:state:Group', group_state - The state of the group, which can be ENABLED, DISABLED, or DELETED.

$sel:disabledDate:Group', group_disabledDate - The date indicating when the group was disabled from Amazon WorkMail use.

$sel:name:Group', group_name - The name of the group.

$sel:id:Group', group_id - The identifier of the group.

$sel:enabledDate:Group', group_enabledDate - The date indicating when the group was enabled for Amazon WorkMail use.

group_email :: Lens' Group (Maybe Text)

The email of the group.

group_state :: Lens' Group (Maybe EntityState)

The state of the group, which can be ENABLED, DISABLED, or DELETED.

group_disabledDate :: Lens' Group (Maybe UTCTime)

The date indicating when the group was disabled from Amazon WorkMail use.

group_name :: Lens' Group (Maybe Text)

The name of the group.

group_id :: Lens' Group (Maybe Text)

The identifier of the group.

group_enabledDate :: Lens' Group (Maybe UTCTime)

The date indicating when the group was enabled for Amazon WorkMail use.

MailDomainSummary

data MailDomainSummary

The data for a given domain.

See: newMailDomainSummary smart constructor.

Constructors

MailDomainSummary'

Fields

defaultDomain :: Maybe Bool

Whether the domain is default or not.

domainName :: Maybe Text

The domain name.

newMailDomainSummary :: MailDomainSummary

Create a value of MailDomainSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultDomain:MailDomainSummary', mailDomainSummary_defaultDomain - Whether the domain is default or not.

$sel:domainName:MailDomainSummary', mailDomainSummary_domainName - The domain name.

mailDomainSummary_defaultDomain :: Lens' MailDomainSummary (Maybe Bool)

Whether the domain is default or not.

mailDomainSummary_domainName :: Lens' MailDomainSummary (Maybe Text)

The domain name.

MailboxExportJob

data MailboxExportJob

The details of a mailbox export job, including the user or resource ID associated with the mailbox and the S3 bucket that the mailbox contents are exported to.

See: newMailboxExportJob smart constructor.

Constructors

MailboxExportJob'

Fields

state :: Maybe MailboxExportJobState

The state of the mailbox export job.

jobId :: Maybe Text

The identifier of the mailbox export job.

startTime :: Maybe POSIX

The mailbox export job start timestamp.

estimatedProgress :: Maybe Natural

The estimated progress of the mailbox export job, in percentage points.

endTime :: Maybe POSIX

The mailbox export job end timestamp.

s3Path :: Maybe Text

The path to the S3 bucket and file that the mailbox export job exports to.

entityId :: Maybe Text

The identifier of the user or resource associated with the mailbox.

description :: Maybe Text

The mailbox export job description.

s3BucketName :: Maybe Text

The name of the S3 bucket.

newMailboxExportJob :: MailboxExportJob

Create a value of MailboxExportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:MailboxExportJob', mailboxExportJob_state - The state of the mailbox export job.

$sel:jobId:MailboxExportJob', mailboxExportJob_jobId - The identifier of the mailbox export job.

$sel:startTime:MailboxExportJob', mailboxExportJob_startTime - The mailbox export job start timestamp.

$sel:estimatedProgress:MailboxExportJob', mailboxExportJob_estimatedProgress - The estimated progress of the mailbox export job, in percentage points.

$sel:endTime:MailboxExportJob', mailboxExportJob_endTime - The mailbox export job end timestamp.

$sel:s3Path:MailboxExportJob', mailboxExportJob_s3Path - The path to the S3 bucket and file that the mailbox export job exports to.

$sel:entityId:MailboxExportJob', mailboxExportJob_entityId - The identifier of the user or resource associated with the mailbox.

$sel:description:MailboxExportJob', mailboxExportJob_description - The mailbox export job description.

$sel:s3BucketName:MailboxExportJob', mailboxExportJob_s3BucketName - The name of the S3 bucket.

mailboxExportJob_state :: Lens' MailboxExportJob (Maybe MailboxExportJobState)

The state of the mailbox export job.

mailboxExportJob_jobId :: Lens' MailboxExportJob (Maybe Text)

The identifier of the mailbox export job.

mailboxExportJob_startTime :: Lens' MailboxExportJob (Maybe UTCTime)

The mailbox export job start timestamp.

mailboxExportJob_estimatedProgress :: Lens' MailboxExportJob (Maybe Natural)

The estimated progress of the mailbox export job, in percentage points.

mailboxExportJob_endTime :: Lens' MailboxExportJob (Maybe UTCTime)

The mailbox export job end timestamp.

mailboxExportJob_s3Path :: Lens' MailboxExportJob (Maybe Text)

The path to the S3 bucket and file that the mailbox export job exports to.

mailboxExportJob_entityId :: Lens' MailboxExportJob (Maybe Text)

The identifier of the user or resource associated with the mailbox.

mailboxExportJob_description :: Lens' MailboxExportJob (Maybe Text)

The mailbox export job description.

mailboxExportJob_s3BucketName :: Lens' MailboxExportJob (Maybe Text)

The name of the S3 bucket.

Member

data Member

The representation of a user or group.

See: newMember smart constructor.

Constructors

Member'

Fields

state :: Maybe EntityState

The state of the member, which can be ENABLED, DISABLED, or DELETED.

disabledDate :: Maybe POSIX

The date indicating when the member was disabled from Amazon WorkMail use.

name :: Maybe Text

The name of the member.

id :: Maybe Text

The identifier of the member.

type' :: Maybe MemberType

A member can be a user or group.

enabledDate :: Maybe POSIX

The date indicating when the member was enabled for Amazon WorkMail use.

newMember :: Member

Create a value of Member with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:Member', member_state - The state of the member, which can be ENABLED, DISABLED, or DELETED.

$sel:disabledDate:Member', member_disabledDate - The date indicating when the member was disabled from Amazon WorkMail use.

$sel:name:Member', member_name - The name of the member.

$sel:id:Member', member_id - The identifier of the member.

$sel:type':Member', member_type - A member can be a user or group.

$sel:enabledDate:Member', member_enabledDate - The date indicating when the member was enabled for Amazon WorkMail use.

member_state :: Lens' Member (Maybe EntityState)

The state of the member, which can be ENABLED, DISABLED, or DELETED.

member_disabledDate :: Lens' Member (Maybe UTCTime)

The date indicating when the member was disabled from Amazon WorkMail use.

member_name :: Lens' Member (Maybe Text)

The name of the member.

member_id :: Lens' Member (Maybe Text)

The identifier of the member.

member_type :: Lens' Member (Maybe MemberType)

A member can be a user or group.

member_enabledDate :: Lens' Member (Maybe UTCTime)

The date indicating when the member was enabled for Amazon WorkMail use.

MobileDeviceAccessMatchedRule

data MobileDeviceAccessMatchedRule

The rule that a simulated user matches.

See: newMobileDeviceAccessMatchedRule smart constructor.

Constructors

MobileDeviceAccessMatchedRule'

Fields

mobileDeviceAccessRuleId :: Maybe Text

Identifier of the rule that a simulated user matches.

name :: Maybe Text

Name of a rule that a simulated user matches.

newMobileDeviceAccessMatchedRule :: MobileDeviceAccessMatchedRule

Create a value of MobileDeviceAccessMatchedRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mobileDeviceAccessRuleId:MobileDeviceAccessMatchedRule', mobileDeviceAccessMatchedRule_mobileDeviceAccessRuleId - Identifier of the rule that a simulated user matches.

$sel:name:MobileDeviceAccessMatchedRule', mobileDeviceAccessMatchedRule_name - Name of a rule that a simulated user matches.

mobileDeviceAccessMatchedRule_mobileDeviceAccessRuleId :: Lens' MobileDeviceAccessMatchedRule (Maybe Text)

Identifier of the rule that a simulated user matches.

mobileDeviceAccessMatchedRule_name :: Lens' MobileDeviceAccessMatchedRule (Maybe Text)

Name of a rule that a simulated user matches.

MobileDeviceAccessOverride

data MobileDeviceAccessOverride

The override object.

See: newMobileDeviceAccessOverride smart constructor.

Constructors

MobileDeviceAccessOverride'

Fields

effect :: Maybe MobileDeviceAccessRuleEffect

The effect of the override, ALLOW or DENY.

userId :: Maybe Text

The WorkMail user to which the access override applies.

dateCreated :: Maybe POSIX

The date the override was first created.

dateModified :: Maybe POSIX

The date the override was last modified.

deviceId :: Maybe Text

The device to which the override applies.

description :: Maybe Text

A description of the override.

newMobileDeviceAccessOverride :: MobileDeviceAccessOverride

Create a value of MobileDeviceAccessOverride with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:effect:MobileDeviceAccessOverride', mobileDeviceAccessOverride_effect - The effect of the override, ALLOW or DENY.

$sel:userId:MobileDeviceAccessOverride', mobileDeviceAccessOverride_userId - The WorkMail user to which the access override applies.

$sel:dateCreated:MobileDeviceAccessOverride', mobileDeviceAccessOverride_dateCreated - The date the override was first created.

$sel:dateModified:MobileDeviceAccessOverride', mobileDeviceAccessOverride_dateModified - The date the override was last modified.

$sel:deviceId:MobileDeviceAccessOverride', mobileDeviceAccessOverride_deviceId - The device to which the override applies.

$sel:description:MobileDeviceAccessOverride', mobileDeviceAccessOverride_description - A description of the override.

mobileDeviceAccessOverride_effect :: Lens' MobileDeviceAccessOverride (Maybe MobileDeviceAccessRuleEffect)

The effect of the override, ALLOW or DENY.

mobileDeviceAccessOverride_userId :: Lens' MobileDeviceAccessOverride (Maybe Text)

The WorkMail user to which the access override applies.

mobileDeviceAccessOverride_dateCreated :: Lens' MobileDeviceAccessOverride (Maybe UTCTime)

The date the override was first created.

mobileDeviceAccessOverride_dateModified :: Lens' MobileDeviceAccessOverride (Maybe UTCTime)

The date the override was last modified.

mobileDeviceAccessOverride_deviceId :: Lens' MobileDeviceAccessOverride (Maybe Text)

The device to which the override applies.

mobileDeviceAccessOverride_description :: Lens' MobileDeviceAccessOverride (Maybe Text)

A description of the override.

MobileDeviceAccessRule

data MobileDeviceAccessRule

A rule that controls access to mobile devices for an Amazon WorkMail group.

See: newMobileDeviceAccessRule smart constructor.

Constructors

MobileDeviceAccessRule'

Fields

effect :: Maybe MobileDeviceAccessRuleEffect

The effect of the rule when it matches. Allowed values are ALLOW or DENY.

deviceUserAgents :: Maybe (NonEmpty Text)

Device user agents that a rule will match.

deviceTypes :: Maybe (NonEmpty Text)

Device types that a rule will match.

notDeviceTypes :: Maybe (NonEmpty Text)

Device types that a rule will not match. All other device types will match.

notDeviceOperatingSystems :: Maybe (NonEmpty Text)

Device operating systems that a rule will not match. All other device types will match.

dateCreated :: Maybe POSIX

The date and time at which an access rule was created.

deviceModels :: Maybe (NonEmpty Text)

Device models that a rule will match.

mobileDeviceAccessRuleId :: Maybe Text

The ID assigned to a mobile access rule.

name :: Maybe Text

The name of a mobile access rule.

dateModified :: Maybe POSIX

The date and time at which an access rule was modified.

deviceOperatingSystems :: Maybe (NonEmpty Text)

Device operating systems that a rule will match.

description :: Maybe Text

The description of a mobile access rule.

notDeviceUserAgents :: Maybe (NonEmpty Text)

Device user agents that a rule will not match. All other device user agents will match.

notDeviceModels :: Maybe (NonEmpty Text)

Device models that a rule will not match. All other device models will match.

newMobileDeviceAccessRule :: MobileDeviceAccessRule

Create a value of MobileDeviceAccessRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:effect:MobileDeviceAccessRule', mobileDeviceAccessRule_effect - The effect of the rule when it matches. Allowed values are ALLOW or DENY.

$sel:deviceUserAgents:MobileDeviceAccessRule', mobileDeviceAccessRule_deviceUserAgents - Device user agents that a rule will match.

$sel:deviceTypes:MobileDeviceAccessRule', mobileDeviceAccessRule_deviceTypes - Device types that a rule will match.

$sel:notDeviceTypes:MobileDeviceAccessRule', mobileDeviceAccessRule_notDeviceTypes - Device types that a rule will not match. All other device types will match.

$sel:notDeviceOperatingSystems:MobileDeviceAccessRule', mobileDeviceAccessRule_notDeviceOperatingSystems - Device operating systems that a rule will not match. All other device types will match.

$sel:dateCreated:MobileDeviceAccessRule', mobileDeviceAccessRule_dateCreated - The date and time at which an access rule was created.

$sel:deviceModels:MobileDeviceAccessRule', mobileDeviceAccessRule_deviceModels - Device models that a rule will match.

$sel:mobileDeviceAccessRuleId:MobileDeviceAccessRule', mobileDeviceAccessRule_mobileDeviceAccessRuleId - The ID assigned to a mobile access rule.

$sel:name:MobileDeviceAccessRule', mobileDeviceAccessRule_name - The name of a mobile access rule.

$sel:dateModified:MobileDeviceAccessRule', mobileDeviceAccessRule_dateModified - The date and time at which an access rule was modified.

$sel:deviceOperatingSystems:MobileDeviceAccessRule', mobileDeviceAccessRule_deviceOperatingSystems - Device operating systems that a rule will match.

$sel:description:MobileDeviceAccessRule', mobileDeviceAccessRule_description - The description of a mobile access rule.

$sel:notDeviceUserAgents:MobileDeviceAccessRule', mobileDeviceAccessRule_notDeviceUserAgents - Device user agents that a rule will not match. All other device user agents will match.

$sel:notDeviceModels:MobileDeviceAccessRule', mobileDeviceAccessRule_notDeviceModels - Device models that a rule will not match. All other device models will match.

mobileDeviceAccessRule_effect :: Lens' MobileDeviceAccessRule (Maybe MobileDeviceAccessRuleEffect)

The effect of the rule when it matches. Allowed values are ALLOW or DENY.

mobileDeviceAccessRule_deviceUserAgents :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device user agents that a rule will match.

mobileDeviceAccessRule_deviceTypes :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device types that a rule will match.

mobileDeviceAccessRule_notDeviceTypes :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device types that a rule will not match. All other device types will match.

mobileDeviceAccessRule_notDeviceOperatingSystems :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device operating systems that a rule will not match. All other device types will match.

mobileDeviceAccessRule_dateCreated :: Lens' MobileDeviceAccessRule (Maybe UTCTime)

The date and time at which an access rule was created.

mobileDeviceAccessRule_deviceModels :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device models that a rule will match.

mobileDeviceAccessRule_mobileDeviceAccessRuleId :: Lens' MobileDeviceAccessRule (Maybe Text)

The ID assigned to a mobile access rule.

mobileDeviceAccessRule_name :: Lens' MobileDeviceAccessRule (Maybe Text)

The name of a mobile access rule.

mobileDeviceAccessRule_dateModified :: Lens' MobileDeviceAccessRule (Maybe UTCTime)

The date and time at which an access rule was modified.

mobileDeviceAccessRule_deviceOperatingSystems :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device operating systems that a rule will match.

mobileDeviceAccessRule_description :: Lens' MobileDeviceAccessRule (Maybe Text)

The description of a mobile access rule.

mobileDeviceAccessRule_notDeviceUserAgents :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device user agents that a rule will not match. All other device user agents will match.

mobileDeviceAccessRule_notDeviceModels :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device models that a rule will not match. All other device models will match.

OrganizationSummary

data OrganizationSummary

The representation of an organization.

See: newOrganizationSummary smart constructor.

Constructors

OrganizationSummary'

Fields

state :: Maybe Text

The state associated with the organization.

alias :: Maybe Text

The alias associated with the organization.

defaultMailDomain :: Maybe Text

The default email domain associated with the organization.

errorMessage :: Maybe Text

The error message associated with the organization. It is only present if unexpected behavior has occurred with regards to the organization. It provides insight or solutions regarding unexpected behavior.

organizationId :: Maybe Text

The identifier associated with the organization.

newOrganizationSummary :: OrganizationSummary

Create a value of OrganizationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:OrganizationSummary', organizationSummary_state - The state associated with the organization.

$sel:alias:OrganizationSummary', organizationSummary_alias - The alias associated with the organization.

$sel:defaultMailDomain:OrganizationSummary', organizationSummary_defaultMailDomain - The default email domain associated with the organization.

$sel:errorMessage:OrganizationSummary', organizationSummary_errorMessage - The error message associated with the organization. It is only present if unexpected behavior has occurred with regards to the organization. It provides insight or solutions regarding unexpected behavior.

$sel:organizationId:OrganizationSummary', organizationSummary_organizationId - The identifier associated with the organization.

organizationSummary_state :: Lens' OrganizationSummary (Maybe Text)

The state associated with the organization.

organizationSummary_alias :: Lens' OrganizationSummary (Maybe Text)

The alias associated with the organization.

organizationSummary_defaultMailDomain :: Lens' OrganizationSummary (Maybe Text)

The default email domain associated with the organization.

organizationSummary_errorMessage :: Lens' OrganizationSummary (Maybe Text)

The error message associated with the organization. It is only present if unexpected behavior has occurred with regards to the organization. It provides insight or solutions regarding unexpected behavior.

organizationSummary_organizationId :: Lens' OrganizationSummary (Maybe Text)

The identifier associated with the organization.

Permission

data Permission

Permission granted to a user, group, or resource to access a certain aspect of another user, group, or resource mailbox.

See: newPermission smart constructor.

Constructors

Permission'

Fields

granteeId :: Text

The identifier of the user, group, or resource to which the permissions are granted.

granteeType :: MemberType

The type of user, group, or resource referred to in GranteeId.

permissionValues :: [PermissionType]

The permissions granted to the grantee. SEND_AS allows the grantee to send email as the owner of the mailbox (the grantee is not mentioned on these emails). SEND_ON_BEHALF allows the grantee to send email on behalf of the owner of the mailbox (the grantee is not mentioned as the physical sender of these emails). FULL_ACCESS allows the grantee full access to the mailbox, irrespective of other folder-level permissions set on the mailbox.

newPermission

Arguments

:: Text
$sel:granteeId:Permission'

-> MemberType
$sel:granteeType:Permission'

-> Permission

Create a value of Permission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:granteeId:Permission', permission_granteeId - The identifier of the user, group, or resource to which the permissions are granted.

$sel:granteeType:Permission', permission_granteeType - The type of user, group, or resource referred to in GranteeId.

$sel:permissionValues:Permission', permission_permissionValues - The permissions granted to the grantee. SEND_AS allows the grantee to send email as the owner of the mailbox (the grantee is not mentioned on these emails). SEND_ON_BEHALF allows the grantee to send email on behalf of the owner of the mailbox (the grantee is not mentioned as the physical sender of these emails). FULL_ACCESS allows the grantee full access to the mailbox, irrespective of other folder-level permissions set on the mailbox.

permission_granteeId :: Lens' Permission Text

The identifier of the user, group, or resource to which the permissions are granted.

permission_granteeType :: Lens' Permission MemberType

The type of user, group, or resource referred to in GranteeId.

permission_permissionValues :: Lens' Permission [PermissionType]

The permissions granted to the grantee. SEND_AS allows the grantee to send email as the owner of the mailbox (the grantee is not mentioned on these emails). SEND_ON_BEHALF allows the grantee to send email on behalf of the owner of the mailbox (the grantee is not mentioned as the physical sender of these emails). FULL_ACCESS allows the grantee full access to the mailbox, irrespective of other folder-level permissions set on the mailbox.

Resource

data Resource

The representation of a resource.

See: newResource smart constructor.

Constructors

Resource'

Fields

email :: Maybe Text

The email of the resource.

state :: Maybe EntityState

The state of the resource, which can be ENABLED, DISABLED, or DELETED.

disabledDate :: Maybe POSIX

The date indicating when the resource was disabled from Amazon WorkMail use.

name :: Maybe Text

The name of the resource.

id :: Maybe Text

The identifier of the resource.

type' :: Maybe ResourceType

The type of the resource: equipment or room.

enabledDate :: Maybe POSIX

The date indicating when the resource was enabled for Amazon WorkMail use.

newResource :: Resource

Create a value of Resource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:email:Resource', resource_email - The email of the resource.

$sel:state:Resource', resource_state - The state of the resource, which can be ENABLED, DISABLED, or DELETED.

$sel:disabledDate:Resource', resource_disabledDate - The date indicating when the resource was disabled from Amazon WorkMail use.

$sel:name:Resource', resource_name - The name of the resource.

$sel:id:Resource', resource_id - The identifier of the resource.

$sel:type':Resource', resource_type - The type of the resource: equipment or room.

$sel:enabledDate:Resource', resource_enabledDate - The date indicating when the resource was enabled for Amazon WorkMail use.

resource_email :: Lens' Resource (Maybe Text)

The email of the resource.

resource_state :: Lens' Resource (Maybe EntityState)

The state of the resource, which can be ENABLED, DISABLED, or DELETED.

resource_disabledDate :: Lens' Resource (Maybe UTCTime)

The date indicating when the resource was disabled from Amazon WorkMail use.

resource_name :: Lens' Resource (Maybe Text)

The name of the resource.

resource_id :: Lens' Resource (Maybe Text)

The identifier of the resource.

resource_type :: Lens' Resource (Maybe ResourceType)

The type of the resource: equipment or room.

resource_enabledDate :: Lens' Resource (Maybe UTCTime)

The date indicating when the resource was enabled for Amazon WorkMail use.

Tag

data Tag

Describes a tag applied to a resource.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

The key of the tag.

value :: Text

The value of the tag.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key of the tag.

$sel:value:Tag', tag_value - The value of the tag.

tag_key :: Lens' Tag Text

The key of the tag.

tag_value :: Lens' Tag Text

The value of the tag.

User

data User

The representation of an Amazon WorkMail user.

See: newUser smart constructor.

Constructors

User'

Fields

email :: Maybe Text

The email of the user.

state :: Maybe EntityState

The state of the user, which can be ENABLED, DISABLED, or DELETED.

disabledDate :: Maybe POSIX

The date indicating when the user was disabled from Amazon WorkMail use.

name :: Maybe Text

The name of the user.

id :: Maybe Text

The identifier of the user.

displayName :: Maybe Text

The display name of the user.

userRole :: Maybe UserRole

The role of the user.

enabledDate :: Maybe POSIX

The date indicating when the user was enabled for Amazon WorkMail use.

newUser :: User

Create a value of User with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:email:User', user_email - The email of the user.

$sel:state:User', user_state - The state of the user, which can be ENABLED, DISABLED, or DELETED.

$sel:disabledDate:User', user_disabledDate - The date indicating when the user was disabled from Amazon WorkMail use.

$sel:name:User', user_name - The name of the user.

$sel:id:User', user_id - The identifier of the user.

$sel:displayName:User', user_displayName - The display name of the user.

$sel:userRole:User', user_userRole - The role of the user.

$sel:enabledDate:User', user_enabledDate - The date indicating when the user was enabled for Amazon WorkMail use.

user_email :: Lens' User (Maybe Text)

The email of the user.

user_state :: Lens' User (Maybe EntityState)

The state of the user, which can be ENABLED, DISABLED, or DELETED.

user_disabledDate :: Lens' User (Maybe UTCTime)

The date indicating when the user was disabled from Amazon WorkMail use.

user_name :: Lens' User (Maybe Text)

The name of the user.

user_id :: Lens' User (Maybe Text)

The identifier of the user.

user_displayName :: Lens' User (Maybe Text)

The display name of the user.

user_userRole :: Lens' User (Maybe UserRole)

The role of the user.

user_enabledDate :: Lens' User (Maybe UTCTime)

The date indicating when the user was enabled for Amazon WorkMail use.