Network.AWS.WorkMail.Types.Delegate

Description

Documentation

data Delegate

The name of the attribute, which is one of the values defined in the UserAttribute enumeration.

See: newDelegate smart constructor.

Constructors

Delegate'

Fields

id :: Text

The identifier for the user or group associated as the resource's delegate.

type' :: MemberType

The type of the delegate: user or group.

newDelegate

Arguments

:: Text
$sel:id:Delegate'

-> MemberType
$sel:type':Delegate'

-> Delegate

Create a value of Delegate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:Delegate', delegate_id - The identifier for the user or group associated as the resource's delegate.

$sel:type':Delegate', delegate_type - The type of the delegate: user or group.

delegate_id :: Lens' Delegate Text

The identifier for the user or group associated as the resource's delegate.

delegate_type :: Lens' Delegate MemberType

The type of the delegate: user or group.