Network.AWS.WorkMail.Types.Domain

Description

Documentation

data Domain

The domain to associate with an Amazon WorkMail organization.

When you configure a domain hosted in Amazon Route 53 (Route 53), all recommended DNS records are added to the organization when you create it. For more information, see Adding a domain in the Amazon WorkMail Administrator Guide.

See: newDomain smart constructor.

Constructors

Domain'

Fields

hostedZoneId :: Maybe Text

The hosted zone ID for a domain hosted in Route 53. Required when configuring a domain hosted in Route 53.

domainName :: Maybe Text

The fully qualified domain name.

newDomain :: Domain

Create a value of Domain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostedZoneId:Domain', domain_hostedZoneId - The hosted zone ID for a domain hosted in Route 53. Required when configuring a domain hosted in Route 53.

$sel:domainName:Domain', domain_domainName - The fully qualified domain name.

domain_hostedZoneId :: Lens' Domain (Maybe Text)

The hosted zone ID for a domain hosted in Route 53. Required when configuring a domain hosted in Route 53.

domain_domainName :: Lens' Domain (Maybe Text)

The fully qualified domain name.