Network.AWS.WorkMail.Types.FolderConfiguration

Description

Documentation

data FolderConfiguration

The configuration applied to an organization's folders by its retention policy.

See: newFolderConfiguration smart constructor.

Constructors

FolderConfiguration'

Fields

period :: Maybe Natural

The period of time at which the folder configuration action is applied.

name :: FolderName

The folder name.

action :: RetentionAction

The action to take on the folder contents at the end of the folder configuration period.

newFolderConfiguration

Arguments

:: FolderName
$sel:name:FolderConfiguration'

-> RetentionAction
$sel:action:FolderConfiguration'

-> FolderConfiguration

Create a value of FolderConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:period:FolderConfiguration', folderConfiguration_period - The period of time at which the folder configuration action is applied.

$sel:name:FolderConfiguration', folderConfiguration_name - The folder name.

$sel:action:FolderConfiguration', folderConfiguration_action - The action to take on the folder contents at the end of the folder configuration period.

folderConfiguration_period :: Lens' FolderConfiguration (Maybe Natural)

The period of time at which the folder configuration action is applied.

folderConfiguration_name :: Lens' FolderConfiguration FolderName

The folder name.

folderConfiguration_action :: Lens' FolderConfiguration RetentionAction

The action to take on the folder contents at the end of the folder configuration period.