Network.AWS.WorkMail.Types.MailboxExportJob

Description

Documentation

data MailboxExportJob

The details of a mailbox export job, including the user or resource ID associated with the mailbox and the S3 bucket that the mailbox contents are exported to.

See: newMailboxExportJob smart constructor.

Constructors

MailboxExportJob'

Fields

estimatedProgress :: Maybe Natural

The estimated progress of the mailbox export job, in percentage points.

entityId :: Maybe Text

The identifier of the user or resource associated with the mailbox.

s3Path :: Maybe Text

The path to the S3 bucket and file that the mailbox export job exports to.

startTime :: Maybe POSIX

The mailbox export job start timestamp.

endTime :: Maybe POSIX

The mailbox export job end timestamp.

state :: Maybe MailboxExportJobState

The state of the mailbox export job.

s3BucketName :: Maybe Text

The name of the S3 bucket.

description :: Maybe Text

The mailbox export job description.

jobId :: Maybe Text

The identifier of the mailbox export job.

newMailboxExportJob :: MailboxExportJob

Create a value of MailboxExportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:estimatedProgress:MailboxExportJob', mailboxExportJob_estimatedProgress - The estimated progress of the mailbox export job, in percentage points.

$sel:entityId:MailboxExportJob', mailboxExportJob_entityId - The identifier of the user or resource associated with the mailbox.

$sel:s3Path:MailboxExportJob', mailboxExportJob_s3Path - The path to the S3 bucket and file that the mailbox export job exports to.

$sel:startTime:MailboxExportJob', mailboxExportJob_startTime - The mailbox export job start timestamp.

$sel:endTime:MailboxExportJob', mailboxExportJob_endTime - The mailbox export job end timestamp.

$sel:state:MailboxExportJob', mailboxExportJob_state - The state of the mailbox export job.

$sel:s3BucketName:MailboxExportJob', mailboxExportJob_s3BucketName - The name of the S3 bucket.

$sel:description:MailboxExportJob', mailboxExportJob_description - The mailbox export job description.

$sel:jobId:MailboxExportJob', mailboxExportJob_jobId - The identifier of the mailbox export job.

mailboxExportJob_estimatedProgress :: Lens' MailboxExportJob (Maybe Natural)

The estimated progress of the mailbox export job, in percentage points.

mailboxExportJob_entityId :: Lens' MailboxExportJob (Maybe Text)

The identifier of the user or resource associated with the mailbox.

mailboxExportJob_s3Path :: Lens' MailboxExportJob (Maybe Text)

The path to the S3 bucket and file that the mailbox export job exports to.

mailboxExportJob_startTime :: Lens' MailboxExportJob (Maybe UTCTime)

The mailbox export job start timestamp.

mailboxExportJob_endTime :: Lens' MailboxExportJob (Maybe UTCTime)

The mailbox export job end timestamp.

mailboxExportJob_state :: Lens' MailboxExportJob (Maybe MailboxExportJobState)

The state of the mailbox export job.

mailboxExportJob_s3BucketName :: Lens' MailboxExportJob (Maybe Text)

The name of the S3 bucket.

mailboxExportJob_description :: Lens' MailboxExportJob (Maybe Text)

The mailbox export job description.

mailboxExportJob_jobId :: Lens' MailboxExportJob (Maybe Text)

The identifier of the mailbox export job.