Network.AWS.WorkMail.Types.MobileDeviceAccessRule

Description

Documentation

data MobileDeviceAccessRule

A rule that controls access to mobile devices for an Amazon WorkMail group.

See: newMobileDeviceAccessRule smart constructor.

Constructors

MobileDeviceAccessRule'

Fields

effect :: Maybe MobileDeviceAccessRuleEffect

The effect of the rule when it matches. Allowed values are ALLOW or DENY.

dateCreated :: Maybe POSIX

The date and time at which an access rule was created.

notDeviceModels :: Maybe (NonEmpty Text)

Device models that a rule will not match. All other device models will match.

notDeviceOperatingSystems :: Maybe (NonEmpty Text)

Device operating systems that a rule will not match. All other device types will match.

deviceUserAgents :: Maybe (NonEmpty Text)

Device user agents that a rule will match.

dateModified :: Maybe POSIX

The date and time at which an access rule was modified.

name :: Maybe Text

The name of a mobile access rule.

mobileDeviceAccessRuleId :: Maybe Text

The ID assigned to a mobile access rule.

deviceModels :: Maybe (NonEmpty Text)

Device models that a rule will match.

notDeviceTypes :: Maybe (NonEmpty Text)

Device types that a rule will not match. All other device types will match.

description :: Maybe Text

The description of a mobile access rule.

notDeviceUserAgents :: Maybe (NonEmpty Text)

Device user agents that a rule will not match. All other device user agents will match.

deviceTypes :: Maybe (NonEmpty Text)

Device types that a rule will match.

deviceOperatingSystems :: Maybe (NonEmpty Text)

Device operating systems that a rule will match.

newMobileDeviceAccessRule :: MobileDeviceAccessRule

Create a value of MobileDeviceAccessRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:effect:MobileDeviceAccessRule', mobileDeviceAccessRule_effect - The effect of the rule when it matches. Allowed values are ALLOW or DENY.

$sel:dateCreated:MobileDeviceAccessRule', mobileDeviceAccessRule_dateCreated - The date and time at which an access rule was created.

$sel:notDeviceModels:MobileDeviceAccessRule', mobileDeviceAccessRule_notDeviceModels - Device models that a rule will not match. All other device models will match.

$sel:notDeviceOperatingSystems:MobileDeviceAccessRule', mobileDeviceAccessRule_notDeviceOperatingSystems - Device operating systems that a rule will not match. All other device types will match.

$sel:deviceUserAgents:MobileDeviceAccessRule', mobileDeviceAccessRule_deviceUserAgents - Device user agents that a rule will match.

$sel:dateModified:MobileDeviceAccessRule', mobileDeviceAccessRule_dateModified - The date and time at which an access rule was modified.

$sel:name:MobileDeviceAccessRule', mobileDeviceAccessRule_name - The name of a mobile access rule.

$sel:mobileDeviceAccessRuleId:MobileDeviceAccessRule', mobileDeviceAccessRule_mobileDeviceAccessRuleId - The ID assigned to a mobile access rule.

$sel:deviceModels:MobileDeviceAccessRule', mobileDeviceAccessRule_deviceModels - Device models that a rule will match.

$sel:notDeviceTypes:MobileDeviceAccessRule', mobileDeviceAccessRule_notDeviceTypes - Device types that a rule will not match. All other device types will match.

$sel:description:MobileDeviceAccessRule', mobileDeviceAccessRule_description - The description of a mobile access rule.

$sel:notDeviceUserAgents:MobileDeviceAccessRule', mobileDeviceAccessRule_notDeviceUserAgents - Device user agents that a rule will not match. All other device user agents will match.

$sel:deviceTypes:MobileDeviceAccessRule', mobileDeviceAccessRule_deviceTypes - Device types that a rule will match.

$sel:deviceOperatingSystems:MobileDeviceAccessRule', mobileDeviceAccessRule_deviceOperatingSystems - Device operating systems that a rule will match.

mobileDeviceAccessRule_effect :: Lens' MobileDeviceAccessRule (Maybe MobileDeviceAccessRuleEffect)

The effect of the rule when it matches. Allowed values are ALLOW or DENY.

mobileDeviceAccessRule_dateCreated :: Lens' MobileDeviceAccessRule (Maybe UTCTime)

The date and time at which an access rule was created.

mobileDeviceAccessRule_notDeviceModels :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device models that a rule will not match. All other device models will match.

mobileDeviceAccessRule_notDeviceOperatingSystems :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device operating systems that a rule will not match. All other device types will match.

mobileDeviceAccessRule_deviceUserAgents :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device user agents that a rule will match.

mobileDeviceAccessRule_dateModified :: Lens' MobileDeviceAccessRule (Maybe UTCTime)

The date and time at which an access rule was modified.

mobileDeviceAccessRule_name :: Lens' MobileDeviceAccessRule (Maybe Text)

The name of a mobile access rule.

mobileDeviceAccessRule_mobileDeviceAccessRuleId :: Lens' MobileDeviceAccessRule (Maybe Text)

The ID assigned to a mobile access rule.

mobileDeviceAccessRule_deviceModels :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device models that a rule will match.

mobileDeviceAccessRule_notDeviceTypes :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device types that a rule will not match. All other device types will match.

mobileDeviceAccessRule_description :: Lens' MobileDeviceAccessRule (Maybe Text)

The description of a mobile access rule.

mobileDeviceAccessRule_notDeviceUserAgents :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device user agents that a rule will not match. All other device user agents will match.

mobileDeviceAccessRule_deviceTypes :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device types that a rule will match.

mobileDeviceAccessRule_deviceOperatingSystems :: Lens' MobileDeviceAccessRule (Maybe (NonEmpty Text))

Device operating systems that a rule will match.