Network.AWS.WorkMail.Types.Permission

Description

Documentation

data Permission

Permission granted to a user, group, or resource to access a certain aspect of another user, group, or resource mailbox.

See: newPermission smart constructor.

Constructors

Permission'

Fields

granteeId :: Text

The identifier of the user, group, or resource to which the permissions are granted.

granteeType :: MemberType

The type of user, group, or resource referred to in GranteeId.

permissionValues :: [PermissionType]

The permissions granted to the grantee. SEND_AS allows the grantee to send email as the owner of the mailbox (the grantee is not mentioned on these emails). SEND_ON_BEHALF allows the grantee to send email on behalf of the owner of the mailbox (the grantee is not mentioned as the physical sender of these emails). FULL_ACCESS allows the grantee full access to the mailbox, irrespective of other folder-level permissions set on the mailbox.

newPermission

Arguments

:: Text
$sel:granteeId:Permission'

-> MemberType
$sel:granteeType:Permission'

-> Permission

Create a value of Permission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:granteeId:Permission', permission_granteeId - The identifier of the user, group, or resource to which the permissions are granted.

$sel:granteeType:Permission', permission_granteeType - The type of user, group, or resource referred to in GranteeId.

$sel:permissionValues:Permission', permission_permissionValues - The permissions granted to the grantee. SEND_AS allows the grantee to send email as the owner of the mailbox (the grantee is not mentioned on these emails). SEND_ON_BEHALF allows the grantee to send email on behalf of the owner of the mailbox (the grantee is not mentioned as the physical sender of these emails). FULL_ACCESS allows the grantee full access to the mailbox, irrespective of other folder-level permissions set on the mailbox.

permission_granteeId :: Lens' Permission Text

The identifier of the user, group, or resource to which the permissions are granted.

permission_granteeType :: Lens' Permission MemberType

The type of user, group, or resource referred to in GranteeId.

permission_permissionValues :: Lens' Permission [PermissionType]

The permissions granted to the grantee. SEND_AS allows the grantee to send email as the owner of the mailbox (the grantee is not mentioned on these emails). SEND_ON_BEHALF allows the grantee to send email on behalf of the owner of the mailbox (the grantee is not mentioned as the physical sender of these emails). FULL_ACCESS allows the grantee full access to the mailbox, irrespective of other folder-level permissions set on the mailbox.