Network.AWS.WorkMail.UpdateMobileDeviceAccessRule

Description

Updates a mobile device access rule for the specified Amazon WorkMail organization.

Creating a Request

data UpdateMobileDeviceAccessRule

See: newUpdateMobileDeviceAccessRule smart constructor.

Constructors

UpdateMobileDeviceAccessRule'

Fields

deviceUserAgents :: Maybe (NonEmpty Text)

User agents that the updated rule will match.

deviceTypes :: Maybe (NonEmpty Text)

Device types that the updated rule will match.

notDeviceTypes :: Maybe (NonEmpty Text)

Device types that the updated rule will not match. All other device types will match.

notDeviceOperatingSystems :: Maybe (NonEmpty Text)

Device operating systems that the updated rule will not match. All other device operating systems will match.

deviceModels :: Maybe (NonEmpty Text)

Device models that the updated rule will match.

deviceOperatingSystems :: Maybe (NonEmpty Text)

Device operating systems that the updated rule will match.

description :: Maybe Text

The updated rule description.

notDeviceUserAgents :: Maybe (NonEmpty Text)

User agents that the updated rule will not match. All other user agents will match.

notDeviceModels :: Maybe (NonEmpty Text)

Device models that the updated rule will not match. All other device models will match.

organizationId :: Text

The Amazon WorkMail organization under which the rule will be updated.

mobileDeviceAccessRuleId :: Text

The identifier of the rule to be updated.

name :: Text

The updated rule name.

effect :: MobileDeviceAccessRuleEffect

The effect of the rule when it matches. Allowed values are ALLOW or DENY.

newUpdateMobileDeviceAccessRule

Arguments

:: Text
$sel:organizationId:UpdateMobileDeviceAccessRule'

-> Text
$sel:mobileDeviceAccessRuleId:UpdateMobileDeviceAccessRule'

-> Text
$sel:name:UpdateMobileDeviceAccessRule'

-> MobileDeviceAccessRuleEffect
$sel:effect:UpdateMobileDeviceAccessRule'

-> UpdateMobileDeviceAccessRule

Create a value of UpdateMobileDeviceAccessRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceUserAgents:UpdateMobileDeviceAccessRule', updateMobileDeviceAccessRule_deviceUserAgents - User agents that the updated rule will match.

$sel:deviceTypes:UpdateMobileDeviceAccessRule', updateMobileDeviceAccessRule_deviceTypes - Device types that the updated rule will match.

$sel:notDeviceTypes:UpdateMobileDeviceAccessRule', updateMobileDeviceAccessRule_notDeviceTypes - Device types that the updated rule will not match. All other device types will match.

$sel:notDeviceOperatingSystems:UpdateMobileDeviceAccessRule', updateMobileDeviceAccessRule_notDeviceOperatingSystems - Device operating systems that the updated rule will not match. All other device operating systems will match.

$sel:deviceModels:UpdateMobileDeviceAccessRule', updateMobileDeviceAccessRule_deviceModels - Device models that the updated rule will match.

$sel:deviceOperatingSystems:UpdateMobileDeviceAccessRule', updateMobileDeviceAccessRule_deviceOperatingSystems - Device operating systems that the updated rule will match.

$sel:description:UpdateMobileDeviceAccessRule', updateMobileDeviceAccessRule_description - The updated rule description.

$sel:notDeviceUserAgents:UpdateMobileDeviceAccessRule', updateMobileDeviceAccessRule_notDeviceUserAgents - User agents that the updated rule will not match. All other user agents will match.

$sel:notDeviceModels:UpdateMobileDeviceAccessRule', updateMobileDeviceAccessRule_notDeviceModels - Device models that the updated rule will not match. All other device models will match.

$sel:organizationId:UpdateMobileDeviceAccessRule', updateMobileDeviceAccessRule_organizationId - The Amazon WorkMail organization under which the rule will be updated.

$sel:mobileDeviceAccessRuleId:UpdateMobileDeviceAccessRule', updateMobileDeviceAccessRule_mobileDeviceAccessRuleId - The identifier of the rule to be updated.

$sel:name:UpdateMobileDeviceAccessRule', updateMobileDeviceAccessRule_name - The updated rule name.

$sel:effect:UpdateMobileDeviceAccessRule', updateMobileDeviceAccessRule_effect - The effect of the rule when it matches. Allowed values are ALLOW or DENY.

Request Lenses

updateMobileDeviceAccessRule_deviceUserAgents :: Lens' UpdateMobileDeviceAccessRule (Maybe (NonEmpty Text))

User agents that the updated rule will match.

updateMobileDeviceAccessRule_deviceTypes :: Lens' UpdateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device types that the updated rule will match.

updateMobileDeviceAccessRule_notDeviceTypes :: Lens' UpdateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device types that the updated rule will not match. All other device types will match.

updateMobileDeviceAccessRule_notDeviceOperatingSystems :: Lens' UpdateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device operating systems that the updated rule will not match. All other device operating systems will match.

updateMobileDeviceAccessRule_deviceModels :: Lens' UpdateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device models that the updated rule will match.

updateMobileDeviceAccessRule_deviceOperatingSystems :: Lens' UpdateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device operating systems that the updated rule will match.

updateMobileDeviceAccessRule_description :: Lens' UpdateMobileDeviceAccessRule (Maybe Text)

The updated rule description.

updateMobileDeviceAccessRule_notDeviceUserAgents :: Lens' UpdateMobileDeviceAccessRule (Maybe (NonEmpty Text))

User agents that the updated rule will not match. All other user agents will match.

updateMobileDeviceAccessRule_notDeviceModels :: Lens' UpdateMobileDeviceAccessRule (Maybe (NonEmpty Text))

Device models that the updated rule will not match. All other device models will match.

updateMobileDeviceAccessRule_organizationId :: Lens' UpdateMobileDeviceAccessRule Text

The Amazon WorkMail organization under which the rule will be updated.

updateMobileDeviceAccessRule_mobileDeviceAccessRuleId :: Lens' UpdateMobileDeviceAccessRule Text

The identifier of the rule to be updated.

updateMobileDeviceAccessRule_name :: Lens' UpdateMobileDeviceAccessRule Text

The updated rule name.

updateMobileDeviceAccessRule_effect :: Lens' UpdateMobileDeviceAccessRule MobileDeviceAccessRuleEffect

The effect of the rule when it matches. Allowed values are ALLOW or DENY.

Destructuring the Response

data UpdateMobileDeviceAccessRuleResponse

See: newUpdateMobileDeviceAccessRuleResponse smart constructor.

Constructors

UpdateMobileDeviceAccessRuleResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateMobileDeviceAccessRuleResponse

Arguments

:: Int
$sel:httpStatus:UpdateMobileDeviceAccessRuleResponse'

-> UpdateMobileDeviceAccessRuleResponse

Create a value of UpdateMobileDeviceAccessRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateMobileDeviceAccessRuleResponse', updateMobileDeviceAccessRuleResponse_httpStatus - The response's http status code.

Response Lenses

updateMobileDeviceAccessRuleResponse_httpStatus :: Lens' UpdateMobileDeviceAccessRuleResponse Int

The response's http status code.