Network.AWS.WorkMailMessageFlow

Description

Derived from API version 2019-05-01 of the AWS service descriptions, licensed under Apache 2.0.

The WorkMail Message Flow API provides access to email messages as they are being sent and received by a WorkMail organization.

Service Configuration

defaultService :: Service

API version 2019-05-01 of the Amazon WorkMail Message Flow SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by WorkMailMessageFlow.

MessageRejected

_MessageRejected :: AsError a => Getting (First ServiceError) a ServiceError

The requested email could not be updated due to an error in the MIME content. Check the error message for more information about what caused the error.

InvalidContentLocation

_InvalidContentLocation :: AsError a => Getting (First ServiceError) a ServiceError

WorkMail could not access the updated email content. Possible reasons:

You made the request in a region other than your S3 bucket region.

The S3 bucket owner is not the same as the calling AWS account.

You have an incomplete or missing S3 bucket policy. For more information about policies, see Updating message content with AWS Lambda in the WorkMail Administrator Guide.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The requested email message is not found.

MessageFrozen

_MessageFrozen :: AsError a => Getting (First ServiceError) a ServiceError

The requested email is not eligible for update. This is usually the case for a redirected email.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

PutRawMessageContent

data PutRawMessageContent

See: newPutRawMessageContent smart constructor.

Constructors

PutRawMessageContent' Text RawMessageContent

newPutRawMessageContent

Arguments

:: Text
$sel:messageId:PutRawMessageContent'

-> RawMessageContent
$sel:content:PutRawMessageContent'

-> PutRawMessageContent

Create a value of PutRawMessageContent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:messageId:PutRawMessageContent', putRawMessageContent_messageId - The identifier of the email message being updated.

$sel:content:PutRawMessageContent', putRawMessageContent_content - Describes the raw message content of the updated email message.

data PutRawMessageContentResponse

See: newPutRawMessageContentResponse smart constructor.

Constructors

PutRawMessageContentResponse' Int

newPutRawMessageContentResponse

Arguments

:: Int
$sel:httpStatus:PutRawMessageContentResponse'

-> PutRawMessageContentResponse

Create a value of PutRawMessageContentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutRawMessageContentResponse', putRawMessageContentResponse_httpStatus - The response's http status code.

GetRawMessageContent

data GetRawMessageContent

See: newGetRawMessageContent smart constructor.

Constructors

GetRawMessageContent' Text

newGetRawMessageContent

Arguments

:: Text
$sel:messageId:GetRawMessageContent'

-> GetRawMessageContent

Create a value of GetRawMessageContent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:messageId:GetRawMessageContent', getRawMessageContent_messageId - The identifier of the email message to retrieve.

data GetRawMessageContentResponse

See: newGetRawMessageContentResponse smart constructor.

Constructors

GetRawMessageContentResponse' Int ResponseBody

newGetRawMessageContentResponse

Arguments

:: Int
$sel:httpStatus:GetRawMessageContentResponse'

-> ResponseBody
$sel:messageContent:GetRawMessageContentResponse'

-> GetRawMessageContentResponse

Create a value of GetRawMessageContentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetRawMessageContentResponse', getRawMessageContentResponse_httpStatus - The response's http status code.

$sel:messageContent:GetRawMessageContentResponse', getRawMessageContentResponse_messageContent - The raw content of the email message, in MIME format.

Types

RawMessageContent

data RawMessageContent

Provides the MIME content of the updated email message as an S3 object. All MIME content must meet the following criteria:

Each part of a multipart MIME message must be formatted properly.

Attachments must be of a content type that Amazon SES supports. For more information, see Unsupported Attachment Types.

If any of the MIME parts in a message contain content that is outside of the 7-bit ASCII character range, we recommend encoding that content.

Per RFC 5321, the maximum length of each line of text, including the <CRLF>, must not exceed 1,000 characters.

The message must contain all the required header fields. Check the returned error message for more information.

The value of immutable headers must remain unchanged. Check the returned error message for more information.

Certain unique headers can only appear once. Check the returned error message for more information.

See: newRawMessageContent smart constructor.

Constructors

RawMessageContent' S3Reference

newRawMessageContent

Arguments

:: S3Reference
$sel:s3Reference:RawMessageContent'

-> RawMessageContent

Create a value of RawMessageContent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Reference:RawMessageContent', rawMessageContent_s3Reference - The S3 reference of an email message.

S3Reference

data S3Reference

Amazon S3 object representing the updated message content, in MIME format.

The region for the S3 bucket containing the S3 object must match the region used for WorkMail operations. Also, for WorkMail to process an S3 object, it must have permission to access that object. For more information, see Updating message content with AWS Lambda.

See: newS3Reference smart constructor.

Constructors

S3Reference' (Maybe Text) Text Text

newS3Reference

Arguments

:: Text
$sel:bucket:S3Reference'

-> Text
$sel:key:S3Reference'

-> S3Reference

Create a value of S3Reference with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:objectVersion:S3Reference', s3Reference_objectVersion - If you enable versioning for the bucket, you can specify the object version.

$sel:bucket:S3Reference', s3Reference_bucket - The S3 bucket name.

$sel:key:S3Reference', s3Reference_key - The S3 key object name.