Network.AWS.WorkMailMessageFlow.Lens

Description

Operations

PutRawMessageContent

putRawMessageContent_messageId :: Lens' PutRawMessageContent Text

The identifier of the email message being updated.

putRawMessageContent_content :: Lens' PutRawMessageContent RawMessageContent

Describes the raw message content of the updated email message.

putRawMessageContentResponse_httpStatus :: Lens' PutRawMessageContentResponse Int

The response's http status code.

GetRawMessageContent

getRawMessageContent_messageId :: Lens' GetRawMessageContent Text

The identifier of the email message to retrieve.

getRawMessageContentResponse_httpStatus :: Lens' GetRawMessageContentResponse Int

The response's http status code.

getRawMessageContentResponse_messageContent :: Lens' GetRawMessageContentResponse ResponseBody

The raw content of the email message, in MIME format.

Types

RawMessageContent

rawMessageContent_s3Reference :: Lens' RawMessageContent S3Reference

The S3 reference of an email message.

S3Reference

s3Reference_objectVersion :: Lens' S3Reference (Maybe Text)

If you enable versioning for the bucket, you can specify the object version.

s3Reference_bucket :: Lens' S3Reference Text

The S3 bucket name.

s3Reference_key :: Lens' S3Reference Text

The S3 key object name.