Network.AWS.WorkSpaces.AssociateConnectionAlias

Description

Associates the specified connection alias with the specified directory to enable cross-Region redirection. For more information, see Cross-Region Redirection for Amazon WorkSpaces.

Before performing this operation, call DescribeConnectionAliases to make sure that the current state of the connection alias is CREATED.

Creating a Request

data AssociateConnectionAlias

See: newAssociateConnectionAlias smart constructor.

Constructors

AssociateConnectionAlias'

Fields

aliasId :: Text

The identifier of the connection alias.

resourceId :: Text

The identifier of the directory to associate the connection alias with.

newAssociateConnectionAlias

Arguments

:: Text
$sel:aliasId:AssociateConnectionAlias'

-> Text
$sel:resourceId:AssociateConnectionAlias'

-> AssociateConnectionAlias

Create a value of AssociateConnectionAlias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aliasId:AssociateConnectionAlias', associateConnectionAlias_aliasId - The identifier of the connection alias.

$sel:resourceId:AssociateConnectionAlias', associateConnectionAlias_resourceId - The identifier of the directory to associate the connection alias with.

Request Lenses

associateConnectionAlias_aliasId :: Lens' AssociateConnectionAlias Text

The identifier of the connection alias.

associateConnectionAlias_resourceId :: Lens' AssociateConnectionAlias Text

The identifier of the directory to associate the connection alias with.

Destructuring the Response

data AssociateConnectionAliasResponse

See: newAssociateConnectionAliasResponse smart constructor.

Constructors

AssociateConnectionAliasResponse'

Fields

connectionIdentifier :: Maybe Text

The identifier of the connection alias association. You use the connection identifier in the DNS TXT record when you're configuring your DNS routing policies.

httpStatus :: Int

The response's http status code.

newAssociateConnectionAliasResponse

Arguments

:: Int
$sel:httpStatus:AssociateConnectionAliasResponse'

-> AssociateConnectionAliasResponse

Create a value of AssociateConnectionAliasResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionIdentifier:AssociateConnectionAliasResponse', associateConnectionAliasResponse_connectionIdentifier - The identifier of the connection alias association. You use the connection identifier in the DNS TXT record when you're configuring your DNS routing policies.

$sel:httpStatus:AssociateConnectionAliasResponse', associateConnectionAliasResponse_httpStatus - The response's http status code.

Response Lenses

associateConnectionAliasResponse_connectionIdentifier :: Lens' AssociateConnectionAliasResponse (Maybe Text)

The identifier of the connection alias association. You use the connection identifier in the DNS TXT record when you're configuring your DNS routing policies.

associateConnectionAliasResponse_httpStatus :: Lens' AssociateConnectionAliasResponse Int

The response's http status code.