Network.AWS.WorkSpaces.DeleteWorkspaceImage

Description

Deletes the specified image from your account. To delete an image, you must first delete any bundles that are associated with the image and unshare the image if it is shared with other accounts.

Creating a Request

data DeleteWorkspaceImage

See: newDeleteWorkspaceImage smart constructor.

Constructors

DeleteWorkspaceImage'

Fields

imageId :: Text

The identifier of the image.

newDeleteWorkspaceImage

Arguments

:: Text
$sel:imageId:DeleteWorkspaceImage'

-> DeleteWorkspaceImage

Create a value of DeleteWorkspaceImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageId:DeleteWorkspaceImage', deleteWorkspaceImage_imageId - The identifier of the image.

Request Lenses

deleteWorkspaceImage_imageId :: Lens' DeleteWorkspaceImage Text

The identifier of the image.

Destructuring the Response

data DeleteWorkspaceImageResponse

See: newDeleteWorkspaceImageResponse smart constructor.

Constructors

DeleteWorkspaceImageResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteWorkspaceImageResponse

Arguments

:: Int
$sel:httpStatus:DeleteWorkspaceImageResponse'

-> DeleteWorkspaceImageResponse

Create a value of DeleteWorkspaceImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteWorkspaceImageResponse', deleteWorkspaceImageResponse_httpStatus - The response's http status code.

Response Lenses

deleteWorkspaceImageResponse_httpStatus :: Lens' DeleteWorkspaceImageResponse Int

The response's http status code.