Network.AWS.WorkSpaces.DescribeConnectionAliasPermissions

Description

Describes the permissions that the owner of a connection alias has granted to another Amazon Web Services account for the specified connection alias. For more information, see Cross-Region Redirection for Amazon WorkSpaces.

Creating a Request

data DescribeConnectionAliasPermissions

See: newDescribeConnectionAliasPermissions smart constructor.

Constructors

DescribeConnectionAliasPermissions'

Fields

nextToken :: Maybe Text

If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return.

aliasId :: Text

The identifier of the connection alias.

newDescribeConnectionAliasPermissions

Arguments

:: Text
$sel:aliasId:DescribeConnectionAliasPermissions'

-> DescribeConnectionAliasPermissions

Create a value of DescribeConnectionAliasPermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeConnectionAliasPermissions', describeConnectionAliasPermissions_nextToken - If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

$sel:maxResults:DescribeConnectionAliasPermissions', describeConnectionAliasPermissions_maxResults - The maximum number of results to return.

$sel:aliasId:DescribeConnectionAliasPermissions', describeConnectionAliasPermissions_aliasId - The identifier of the connection alias.

Request Lenses

describeConnectionAliasPermissions_nextToken :: Lens' DescribeConnectionAliasPermissions (Maybe Text)

If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

describeConnectionAliasPermissions_maxResults :: Lens' DescribeConnectionAliasPermissions (Maybe Natural)

The maximum number of results to return.

describeConnectionAliasPermissions_aliasId :: Lens' DescribeConnectionAliasPermissions Text

The identifier of the connection alias.

Destructuring the Response

data DescribeConnectionAliasPermissionsResponse

See: newDescribeConnectionAliasPermissionsResponse smart constructor.

Constructors

DescribeConnectionAliasPermissionsResponse'

Fields

aliasId :: Maybe Text

The identifier of the connection alias.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

connectionAliasPermissions :: Maybe (NonEmpty ConnectionAliasPermission)

The permissions associated with a connection alias.

httpStatus :: Int

The response's http status code.

newDescribeConnectionAliasPermissionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeConnectionAliasPermissionsResponse'

-> DescribeConnectionAliasPermissionsResponse

Create a value of DescribeConnectionAliasPermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aliasId:DescribeConnectionAliasPermissions', describeConnectionAliasPermissionsResponse_aliasId - The identifier of the connection alias.

$sel:nextToken:DescribeConnectionAliasPermissions', describeConnectionAliasPermissionsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:connectionAliasPermissions:DescribeConnectionAliasPermissionsResponse', describeConnectionAliasPermissionsResponse_connectionAliasPermissions - The permissions associated with a connection alias.

$sel:httpStatus:DescribeConnectionAliasPermissionsResponse', describeConnectionAliasPermissionsResponse_httpStatus - The response's http status code.

Response Lenses

describeConnectionAliasPermissionsResponse_aliasId :: Lens' DescribeConnectionAliasPermissionsResponse (Maybe Text)

The identifier of the connection alias.

describeConnectionAliasPermissionsResponse_nextToken :: Lens' DescribeConnectionAliasPermissionsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeConnectionAliasPermissionsResponse_connectionAliasPermissions :: Lens' DescribeConnectionAliasPermissionsResponse (Maybe (NonEmpty ConnectionAliasPermission))

The permissions associated with a connection alias.

describeConnectionAliasPermissionsResponse_httpStatus :: Lens' DescribeConnectionAliasPermissionsResponse Int

The response's http status code.