Network.AWS.WorkSpaces.DescribeConnectionAliases

Description

Retrieves a list that describes the connection aliases used for cross-Region redirection. For more information, see Cross-Region Redirection for Amazon WorkSpaces.

Creating a Request

data DescribeConnectionAliases

See: newDescribeConnectionAliases smart constructor.

Constructors

DescribeConnectionAliases'

Fields

resourceId :: Maybe Text

The identifier of the directory associated with the connection alias.

nextToken :: Maybe Text

If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

aliasIds :: Maybe (NonEmpty Text)

The identifiers of the connection aliases to describe.

limit :: Maybe Natural

The maximum number of connection aliases to return.

newDescribeConnectionAliases :: DescribeConnectionAliases

Create a value of DescribeConnectionAliases with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:DescribeConnectionAliases', describeConnectionAliases_resourceId - The identifier of the directory associated with the connection alias.

$sel:nextToken:DescribeConnectionAliases', describeConnectionAliases_nextToken - If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

$sel:aliasIds:DescribeConnectionAliases', describeConnectionAliases_aliasIds - The identifiers of the connection aliases to describe.

$sel:limit:DescribeConnectionAliases', describeConnectionAliases_limit - The maximum number of connection aliases to return.

Request Lenses

describeConnectionAliases_resourceId :: Lens' DescribeConnectionAliases (Maybe Text)

The identifier of the directory associated with the connection alias.

describeConnectionAliases_nextToken :: Lens' DescribeConnectionAliases (Maybe Text)

If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

describeConnectionAliases_aliasIds :: Lens' DescribeConnectionAliases (Maybe (NonEmpty Text))

The identifiers of the connection aliases to describe.

describeConnectionAliases_limit :: Lens' DescribeConnectionAliases (Maybe Natural)

The maximum number of connection aliases to return.

Destructuring the Response

data DescribeConnectionAliasesResponse

See: newDescribeConnectionAliasesResponse smart constructor.

Constructors

DescribeConnectionAliasesResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

connectionAliases :: Maybe (NonEmpty ConnectionAlias)

Information about the specified connection aliases.

httpStatus :: Int

The response's http status code.

newDescribeConnectionAliasesResponse

Arguments

:: Int
$sel:httpStatus:DescribeConnectionAliasesResponse'

-> DescribeConnectionAliasesResponse

Create a value of DescribeConnectionAliasesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeConnectionAliases', describeConnectionAliasesResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:connectionAliases:DescribeConnectionAliasesResponse', describeConnectionAliasesResponse_connectionAliases - Information about the specified connection aliases.

$sel:httpStatus:DescribeConnectionAliasesResponse', describeConnectionAliasesResponse_httpStatus - The response's http status code.

Response Lenses

describeConnectionAliasesResponse_nextToken :: Lens' DescribeConnectionAliasesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeConnectionAliasesResponse_connectionAliases :: Lens' DescribeConnectionAliasesResponse (Maybe (NonEmpty ConnectionAlias))

Information about the specified connection aliases.

describeConnectionAliasesResponse_httpStatus :: Lens' DescribeConnectionAliasesResponse Int

The response's http status code.