Network.AWS.WorkSpaces.DescribeIpGroups

Description

Describes one or more of your IP access control groups.

This operation returns paginated results.

Creating a Request

data DescribeIpGroups

See: newDescribeIpGroups smart constructor.

Constructors

DescribeIpGroups'

Fields

groupIds :: Maybe [Text]

The identifiers of one or more IP access control groups.

nextToken :: Maybe Text

If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

maxResults :: Maybe Natural

The maximum number of items to return.

newDescribeIpGroups :: DescribeIpGroups

Create a value of DescribeIpGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupIds:DescribeIpGroups', describeIpGroups_groupIds - The identifiers of one or more IP access control groups.

$sel:nextToken:DescribeIpGroups', describeIpGroups_nextToken - If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

$sel:maxResults:DescribeIpGroups', describeIpGroups_maxResults - The maximum number of items to return.

Request Lenses

describeIpGroups_groupIds :: Lens' DescribeIpGroups (Maybe [Text])

The identifiers of one or more IP access control groups.

describeIpGroups_nextToken :: Lens' DescribeIpGroups (Maybe Text)

If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

describeIpGroups_maxResults :: Lens' DescribeIpGroups (Maybe Natural)

The maximum number of items to return.

Destructuring the Response

data DescribeIpGroupsResponse

See: newDescribeIpGroupsResponse smart constructor.

Constructors

DescribeIpGroupsResponse'

Fields

result :: Maybe [WorkspacesIpGroup]

Information about the IP access control groups.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeIpGroupsResponse

Arguments

:: Int
$sel:httpStatus:DescribeIpGroupsResponse'

-> DescribeIpGroupsResponse

Create a value of DescribeIpGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:result:DescribeIpGroupsResponse', describeIpGroupsResponse_result - Information about the IP access control groups.

$sel:nextToken:DescribeIpGroups', describeIpGroupsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeIpGroupsResponse', describeIpGroupsResponse_httpStatus - The response's http status code.

Response Lenses

describeIpGroupsResponse_result :: Lens' DescribeIpGroupsResponse (Maybe [WorkspacesIpGroup])

Information about the IP access control groups.

describeIpGroupsResponse_nextToken :: Lens' DescribeIpGroupsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeIpGroupsResponse_httpStatus :: Lens' DescribeIpGroupsResponse Int

The response's http status code.