Network.AWS.WorkSpaces.DescribeTags

Description

Describes the specified tags for the specified WorkSpaces resource.

Creating a Request

data DescribeTags

See: newDescribeTags smart constructor.

Constructors

DescribeTags'

Fields

resourceId :: Text

The identifier of the WorkSpaces resource. The supported resource types are WorkSpaces, registered directories, images, custom bundles, IP access control groups, and connection aliases.

newDescribeTags

Arguments

:: Text
$sel:resourceId:DescribeTags'

-> DescribeTags

Create a value of DescribeTags with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:DescribeTags', describeTags_resourceId - The identifier of the WorkSpaces resource. The supported resource types are WorkSpaces, registered directories, images, custom bundles, IP access control groups, and connection aliases.

Request Lenses

describeTags_resourceId :: Lens' DescribeTags Text

The identifier of the WorkSpaces resource. The supported resource types are WorkSpaces, registered directories, images, custom bundles, IP access control groups, and connection aliases.

Destructuring the Response

data DescribeTagsResponse

See: newDescribeTagsResponse smart constructor.

Constructors

DescribeTagsResponse'

Fields

tagList :: Maybe [Tag]

The tags.

httpStatus :: Int

The response's http status code.

newDescribeTagsResponse

Arguments

:: Int
$sel:httpStatus:DescribeTagsResponse'

-> DescribeTagsResponse

Create a value of DescribeTagsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagList:DescribeTagsResponse', describeTagsResponse_tagList - The tags.

$sel:httpStatus:DescribeTagsResponse', describeTagsResponse_httpStatus - The response's http status code.

Response Lenses

describeTagsResponse_tagList :: Lens' DescribeTagsResponse (Maybe [Tag])

The tags.

describeTagsResponse_httpStatus :: Lens' DescribeTagsResponse Int

The response's http status code.