Network.AWS.WorkSpaces.DescribeWorkspaceImagePermissions

Description

Describes the permissions that the owner of an image has granted to other AWS accounts for an image.

Creating a Request

data DescribeWorkspaceImagePermissions

See: newDescribeWorkspaceImagePermissions smart constructor.

Constructors

DescribeWorkspaceImagePermissions'

Fields

nextToken :: Maybe Text

If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

maxResults :: Maybe Natural

The maximum number of items to return.

imageId :: Text

The identifier of the image.

newDescribeWorkspaceImagePermissions

Arguments

:: Text
$sel:imageId:DescribeWorkspaceImagePermissions'

-> DescribeWorkspaceImagePermissions

Create a value of DescribeWorkspaceImagePermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeWorkspaceImagePermissions', describeWorkspaceImagePermissions_nextToken - If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

$sel:maxResults:DescribeWorkspaceImagePermissions', describeWorkspaceImagePermissions_maxResults - The maximum number of items to return.

$sel:imageId:DescribeWorkspaceImagePermissions', describeWorkspaceImagePermissions_imageId - The identifier of the image.

Request Lenses

describeWorkspaceImagePermissions_nextToken :: Lens' DescribeWorkspaceImagePermissions (Maybe Text)

If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

describeWorkspaceImagePermissions_maxResults :: Lens' DescribeWorkspaceImagePermissions (Maybe Natural)

The maximum number of items to return.

describeWorkspaceImagePermissions_imageId :: Lens' DescribeWorkspaceImagePermissions Text

The identifier of the image.

Destructuring the Response

data DescribeWorkspaceImagePermissionsResponse

See: newDescribeWorkspaceImagePermissionsResponse smart constructor.

Constructors

DescribeWorkspaceImagePermissionsResponse'

Fields

imagePermissions :: Maybe [ImagePermission]

The identifiers of the AWS accounts that the image has been shared with.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

imageId :: Maybe Text

The identifier of the image.

httpStatus :: Int

The response's http status code.

newDescribeWorkspaceImagePermissionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeWorkspaceImagePermissionsResponse'

-> DescribeWorkspaceImagePermissionsResponse

Create a value of DescribeWorkspaceImagePermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imagePermissions:DescribeWorkspaceImagePermissionsResponse', describeWorkspaceImagePermissionsResponse_imagePermissions - The identifiers of the AWS accounts that the image has been shared with.

$sel:nextToken:DescribeWorkspaceImagePermissions', describeWorkspaceImagePermissionsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:imageId:DescribeWorkspaceImagePermissions', describeWorkspaceImagePermissionsResponse_imageId - The identifier of the image.

$sel:httpStatus:DescribeWorkspaceImagePermissionsResponse', describeWorkspaceImagePermissionsResponse_httpStatus - The response's http status code.

Response Lenses

describeWorkspaceImagePermissionsResponse_imagePermissions :: Lens' DescribeWorkspaceImagePermissionsResponse (Maybe [ImagePermission])

The identifiers of the AWS accounts that the image has been shared with.

describeWorkspaceImagePermissionsResponse_nextToken :: Lens' DescribeWorkspaceImagePermissionsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeWorkspaceImagePermissionsResponse_imageId :: Lens' DescribeWorkspaceImagePermissionsResponse (Maybe Text)

The identifier of the image.

describeWorkspaceImagePermissionsResponse_httpStatus :: Lens' DescribeWorkspaceImagePermissionsResponse Int

The response's http status code.