Network.AWS.WorkSpaces.DescribeWorkspaceImages

Description

Retrieves a list that describes one or more specified images, if the image identifiers are provided. Otherwise, all images in the account are described.

This operation returns paginated results.

Creating a Request

data DescribeWorkspaceImages

See: newDescribeWorkspaceImages smart constructor.

Constructors

DescribeWorkspaceImages'

Fields

imageType :: Maybe ImageType

The type (owned or shared) of the image.

nextToken :: Maybe Text

If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

imageIds :: Maybe (NonEmpty Text)

The identifier of the image.

maxResults :: Maybe Natural

The maximum number of items to return.

newDescribeWorkspaceImages :: DescribeWorkspaceImages

Create a value of DescribeWorkspaceImages with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageType:DescribeWorkspaceImages', describeWorkspaceImages_imageType - The type (owned or shared) of the image.

$sel:nextToken:DescribeWorkspaceImages', describeWorkspaceImages_nextToken - If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

$sel:imageIds:DescribeWorkspaceImages', describeWorkspaceImages_imageIds - The identifier of the image.

$sel:maxResults:DescribeWorkspaceImages', describeWorkspaceImages_maxResults - The maximum number of items to return.

Request Lenses

describeWorkspaceImages_imageType :: Lens' DescribeWorkspaceImages (Maybe ImageType)

The type (owned or shared) of the image.

describeWorkspaceImages_nextToken :: Lens' DescribeWorkspaceImages (Maybe Text)

If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

describeWorkspaceImages_imageIds :: Lens' DescribeWorkspaceImages (Maybe (NonEmpty Text))

The identifier of the image.

describeWorkspaceImages_maxResults :: Lens' DescribeWorkspaceImages (Maybe Natural)

The maximum number of items to return.

Destructuring the Response

data DescribeWorkspaceImagesResponse

See: newDescribeWorkspaceImagesResponse smart constructor.

Constructors

DescribeWorkspaceImagesResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

images :: Maybe [WorkspaceImage]

Information about the images.

httpStatus :: Int

The response's http status code.

newDescribeWorkspaceImagesResponse

Arguments

:: Int
$sel:httpStatus:DescribeWorkspaceImagesResponse'

-> DescribeWorkspaceImagesResponse

Create a value of DescribeWorkspaceImagesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeWorkspaceImages', describeWorkspaceImagesResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:images:DescribeWorkspaceImagesResponse', describeWorkspaceImagesResponse_images - Information about the images.

$sel:httpStatus:DescribeWorkspaceImagesResponse', describeWorkspaceImagesResponse_httpStatus - The response's http status code.

Response Lenses

describeWorkspaceImagesResponse_nextToken :: Lens' DescribeWorkspaceImagesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeWorkspaceImagesResponse_images :: Lens' DescribeWorkspaceImagesResponse (Maybe [WorkspaceImage])

Information about the images.

describeWorkspaceImagesResponse_httpStatus :: Lens' DescribeWorkspaceImagesResponse Int

The response's http status code.