Network.AWS.WorkSpaces.ImportWorkspaceImage

Description

Imports the specified Windows 10 Bring Your Own License (BYOL) image into Amazon WorkSpaces. The image must be an already licensed Amazon EC2 image that is in your Amazon Web Services account, and you must own the image. For more information about creating BYOL images, see Bring Your Own Windows Desktop Licenses.

Creating a Request

data ImportWorkspaceImage

See: newImportWorkspaceImage smart constructor.

Constructors

ImportWorkspaceImage'

Fields

applications :: Maybe (NonEmpty Application)

If specified, the version of Microsoft Office to subscribe to. Valid only for Windows 10 BYOL images. For more information about subscribing to Office for BYOL images, see Bring Your Own Windows Desktop Licenses.

Although this parameter is an array, only one item is allowed at this time.

tags :: Maybe [Tag]

The tags. Each WorkSpaces resource can have a maximum of 50 tags.

ec2ImageId :: Text

The identifier of the EC2 image.

ingestionProcess :: WorkspaceImageIngestionProcess

The ingestion process to be used when importing the image, depending on which protocol you want to use for your BYOL Workspace image, either PCoIP or WorkSpaces Streaming Protocol (WSP). To use WSP, specify a value that ends in _WSP. To use PCoIP, specify a value that does not end in _WSP.

For non-GPU-enabled bundles (bundles other than Graphics or GraphicsPro), specify BYOL_REGULAR or BYOL_REGULAR_WSP, depending on the protocol.

imageName :: Text

The name of the WorkSpace image.

imageDescription :: Text

The description of the WorkSpace image.

newImportWorkspaceImage

Arguments

:: Text
$sel:ec2ImageId:ImportWorkspaceImage'

-> WorkspaceImageIngestionProcess
$sel:ingestionProcess:ImportWorkspaceImage'

-> Text
$sel:imageName:ImportWorkspaceImage'

-> Text
$sel:imageDescription:ImportWorkspaceImage'

-> ImportWorkspaceImage

Create a value of ImportWorkspaceImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applications:ImportWorkspaceImage', importWorkspaceImage_applications - If specified, the version of Microsoft Office to subscribe to. Valid only for Windows 10 BYOL images. For more information about subscribing to Office for BYOL images, see Bring Your Own Windows Desktop Licenses.

Although this parameter is an array, only one item is allowed at this time.

$sel:tags:ImportWorkspaceImage', importWorkspaceImage_tags - The tags. Each WorkSpaces resource can have a maximum of 50 tags.

$sel:ec2ImageId:ImportWorkspaceImage', importWorkspaceImage_ec2ImageId - The identifier of the EC2 image.

$sel:ingestionProcess:ImportWorkspaceImage', importWorkspaceImage_ingestionProcess - The ingestion process to be used when importing the image, depending on which protocol you want to use for your BYOL Workspace image, either PCoIP or WorkSpaces Streaming Protocol (WSP). To use WSP, specify a value that ends in _WSP. To use PCoIP, specify a value that does not end in _WSP.

For non-GPU-enabled bundles (bundles other than Graphics or GraphicsPro), specify BYOL_REGULAR or BYOL_REGULAR_WSP, depending on the protocol.

$sel:imageName:ImportWorkspaceImage', importWorkspaceImage_imageName - The name of the WorkSpace image.

$sel:imageDescription:ImportWorkspaceImage', importWorkspaceImage_imageDescription - The description of the WorkSpace image.

Request Lenses

importWorkspaceImage_applications :: Lens' ImportWorkspaceImage (Maybe (NonEmpty Application))

If specified, the version of Microsoft Office to subscribe to. Valid only for Windows 10 BYOL images. For more information about subscribing to Office for BYOL images, see Bring Your Own Windows Desktop Licenses.

Although this parameter is an array, only one item is allowed at this time.

importWorkspaceImage_tags :: Lens' ImportWorkspaceImage (Maybe [Tag])

The tags. Each WorkSpaces resource can have a maximum of 50 tags.

importWorkspaceImage_ec2ImageId :: Lens' ImportWorkspaceImage Text

The identifier of the EC2 image.

importWorkspaceImage_ingestionProcess :: Lens' ImportWorkspaceImage WorkspaceImageIngestionProcess

The ingestion process to be used when importing the image, depending on which protocol you want to use for your BYOL Workspace image, either PCoIP or WorkSpaces Streaming Protocol (WSP). To use WSP, specify a value that ends in _WSP. To use PCoIP, specify a value that does not end in _WSP.

For non-GPU-enabled bundles (bundles other than Graphics or GraphicsPro), specify BYOL_REGULAR or BYOL_REGULAR_WSP, depending on the protocol.

importWorkspaceImage_imageName :: Lens' ImportWorkspaceImage Text

The name of the WorkSpace image.

importWorkspaceImage_imageDescription :: Lens' ImportWorkspaceImage Text

The description of the WorkSpace image.

Destructuring the Response

data ImportWorkspaceImageResponse

See: newImportWorkspaceImageResponse smart constructor.

Constructors

ImportWorkspaceImageResponse'

Fields

imageId :: Maybe Text

The identifier of the WorkSpace image.

httpStatus :: Int

The response's http status code.

newImportWorkspaceImageResponse

Arguments

:: Int
$sel:httpStatus:ImportWorkspaceImageResponse'

-> ImportWorkspaceImageResponse

Create a value of ImportWorkspaceImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageId:ImportWorkspaceImageResponse', importWorkspaceImageResponse_imageId - The identifier of the WorkSpace image.

$sel:httpStatus:ImportWorkspaceImageResponse', importWorkspaceImageResponse_httpStatus - The response's http status code.

Response Lenses

importWorkspaceImageResponse_imageId :: Lens' ImportWorkspaceImageResponse (Maybe Text)

The identifier of the WorkSpace image.

importWorkspaceImageResponse_httpStatus :: Lens' ImportWorkspaceImageResponse Int

The response's http status code.