Network.AWS.WorkSpaces.MigrateWorkspace

Description

Migrates a WorkSpace from one operating system or bundle type to another, while retaining the data on the user volume.

The migration process recreates the WorkSpace by using a new root volume from the target bundle image and the user volume from the last available snapshot of the original WorkSpace. During migration, the original D:\Users\%USERNAME% user profile folder is renamed to D:\Users\%USERNAME%MMddyyTHHmmss%.NotMigrated. A new D:\Users\%USERNAME%\ folder is generated by the new OS. Certain files in the old user profile are moved to the new user profile.

For available migration scenarios, details about what happens during migration, and best practices, see Migrate a WorkSpace.

Creating a Request

data MigrateWorkspace

See: newMigrateWorkspace smart constructor.

Constructors

MigrateWorkspace'

Fields

sourceWorkspaceId :: Text

The identifier of the WorkSpace to migrate from.

bundleId :: Text

The identifier of the target bundle type to migrate the WorkSpace to.

newMigrateWorkspace

Arguments

:: Text
$sel:sourceWorkspaceId:MigrateWorkspace'

-> Text
$sel:bundleId:MigrateWorkspace'

-> MigrateWorkspace

Create a value of MigrateWorkspace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceWorkspaceId:MigrateWorkspace', migrateWorkspace_sourceWorkspaceId - The identifier of the WorkSpace to migrate from.

$sel:bundleId:MigrateWorkspace', migrateWorkspace_bundleId - The identifier of the target bundle type to migrate the WorkSpace to.

Request Lenses

migrateWorkspace_sourceWorkspaceId :: Lens' MigrateWorkspace Text

The identifier of the WorkSpace to migrate from.

migrateWorkspace_bundleId :: Lens' MigrateWorkspace Text

The identifier of the target bundle type to migrate the WorkSpace to.

Destructuring the Response

data MigrateWorkspaceResponse

See: newMigrateWorkspaceResponse smart constructor.

Constructors

MigrateWorkspaceResponse'

Fields

targetWorkspaceId :: Maybe Text

The new identifier of the WorkSpace that is being migrated. If the migration does not succeed, the target WorkSpace ID will not be used, and the WorkSpace will still have the original WorkSpace ID.

sourceWorkspaceId :: Maybe Text

The original identifier of the WorkSpace that is being migrated.

httpStatus :: Int

The response's http status code.

newMigrateWorkspaceResponse

Arguments

:: Int
$sel:httpStatus:MigrateWorkspaceResponse'

-> MigrateWorkspaceResponse

Create a value of MigrateWorkspaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetWorkspaceId:MigrateWorkspaceResponse', migrateWorkspaceResponse_targetWorkspaceId - The new identifier of the WorkSpace that is being migrated. If the migration does not succeed, the target WorkSpace ID will not be used, and the WorkSpace will still have the original WorkSpace ID.

$sel:sourceWorkspaceId:MigrateWorkspace', migrateWorkspaceResponse_sourceWorkspaceId - The original identifier of the WorkSpace that is being migrated.

$sel:httpStatus:MigrateWorkspaceResponse', migrateWorkspaceResponse_httpStatus - The response's http status code.

Response Lenses

migrateWorkspaceResponse_targetWorkspaceId :: Lens' MigrateWorkspaceResponse (Maybe Text)

The new identifier of the WorkSpace that is being migrated. If the migration does not succeed, the target WorkSpace ID will not be used, and the WorkSpace will still have the original WorkSpace ID.

migrateWorkspaceResponse_sourceWorkspaceId :: Lens' MigrateWorkspaceResponse (Maybe Text)

The original identifier of the WorkSpace that is being migrated.

migrateWorkspaceResponse_httpStatus :: Lens' MigrateWorkspaceResponse Int

The response's http status code.