Network.AWS.WorkSpaces.RegisterWorkspaceDirectory

Description

Registers the specified directory. This operation is asynchronous and returns before the WorkSpace directory is registered. If this is the first time you are registering a directory, you will need to create the workspaces_DefaultRole role before you can register a directory. For more information, see Creating the workspaces_DefaultRole Role.

Creating a Request

data RegisterWorkspaceDirectory

See: newRegisterWorkspaceDirectory smart constructor.

Constructors

RegisterWorkspaceDirectory'

Fields

subnetIds :: Maybe [Text]

The identifiers of the subnets for your virtual private cloud (VPC). Make sure that the subnets are in supported Availability Zones. The subnets must also be in separate Availability Zones. If these conditions are not met, you will receive an OperationNotSupportedException error.

tenancy :: Maybe Tenancy

Indicates whether your WorkSpace directory is dedicated or shared. To use Bring Your Own License (BYOL) images, this value must be set to DEDICATED and your AWS account must be enabled for BYOL. If your account has not been enabled for BYOL, you will receive an InvalidParameterValuesException error. For more information about BYOL images, see Bring Your Own Windows Desktop Images.

tags :: Maybe [Tag]

The tags associated with the directory.

enableSelfService :: Maybe Bool

Indicates whether self-service capabilities are enabled or disabled.

directoryId :: Text

The identifier of the directory. You cannot register a directory if it does not have a status of Active. If the directory does not have a status of Active, you will receive an InvalidResourceStateException error. If you have already registered the maximum number of directories that you can register with Amazon WorkSpaces, you will receive a ResourceLimitExceededException error. Deregister directories that you are not using for WorkSpaces, and try again.

enableWorkDocs :: Bool

Indicates whether Amazon WorkDocs is enabled or disabled. If you have enabled this parameter and WorkDocs is not available in the Region, you will receive an OperationNotSupportedException error. Set EnableWorkDocs to disabled, and try again.

newRegisterWorkspaceDirectory

Arguments

:: Text
$sel:directoryId:RegisterWorkspaceDirectory'

-> Bool
$sel:enableWorkDocs:RegisterWorkspaceDirectory'

-> RegisterWorkspaceDirectory

Create a value of RegisterWorkspaceDirectory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subnetIds:RegisterWorkspaceDirectory', registerWorkspaceDirectory_subnetIds - The identifiers of the subnets for your virtual private cloud (VPC). Make sure that the subnets are in supported Availability Zones. The subnets must also be in separate Availability Zones. If these conditions are not met, you will receive an OperationNotSupportedException error.

$sel:tenancy:RegisterWorkspaceDirectory', registerWorkspaceDirectory_tenancy - Indicates whether your WorkSpace directory is dedicated or shared. To use Bring Your Own License (BYOL) images, this value must be set to DEDICATED and your AWS account must be enabled for BYOL. If your account has not been enabled for BYOL, you will receive an InvalidParameterValuesException error. For more information about BYOL images, see Bring Your Own Windows Desktop Images.

$sel:tags:RegisterWorkspaceDirectory', registerWorkspaceDirectory_tags - The tags associated with the directory.

$sel:enableSelfService:RegisterWorkspaceDirectory', registerWorkspaceDirectory_enableSelfService - Indicates whether self-service capabilities are enabled or disabled.

$sel:directoryId:RegisterWorkspaceDirectory', registerWorkspaceDirectory_directoryId - The identifier of the directory. You cannot register a directory if it does not have a status of Active. If the directory does not have a status of Active, you will receive an InvalidResourceStateException error. If you have already registered the maximum number of directories that you can register with Amazon WorkSpaces, you will receive a ResourceLimitExceededException error. Deregister directories that you are not using for WorkSpaces, and try again.

$sel:enableWorkDocs:RegisterWorkspaceDirectory', registerWorkspaceDirectory_enableWorkDocs - Indicates whether Amazon WorkDocs is enabled or disabled. If you have enabled this parameter and WorkDocs is not available in the Region, you will receive an OperationNotSupportedException error. Set EnableWorkDocs to disabled, and try again.

Request Lenses

registerWorkspaceDirectory_subnetIds :: Lens' RegisterWorkspaceDirectory (Maybe [Text])

The identifiers of the subnets for your virtual private cloud (VPC). Make sure that the subnets are in supported Availability Zones. The subnets must also be in separate Availability Zones. If these conditions are not met, you will receive an OperationNotSupportedException error.

registerWorkspaceDirectory_tenancy :: Lens' RegisterWorkspaceDirectory (Maybe Tenancy)

Indicates whether your WorkSpace directory is dedicated or shared. To use Bring Your Own License (BYOL) images, this value must be set to DEDICATED and your AWS account must be enabled for BYOL. If your account has not been enabled for BYOL, you will receive an InvalidParameterValuesException error. For more information about BYOL images, see Bring Your Own Windows Desktop Images.

registerWorkspaceDirectory_tags :: Lens' RegisterWorkspaceDirectory (Maybe [Tag])

The tags associated with the directory.

registerWorkspaceDirectory_enableSelfService :: Lens' RegisterWorkspaceDirectory (Maybe Bool)

Indicates whether self-service capabilities are enabled or disabled.

registerWorkspaceDirectory_directoryId :: Lens' RegisterWorkspaceDirectory Text

The identifier of the directory. You cannot register a directory if it does not have a status of Active. If the directory does not have a status of Active, you will receive an InvalidResourceStateException error. If you have already registered the maximum number of directories that you can register with Amazon WorkSpaces, you will receive a ResourceLimitExceededException error. Deregister directories that you are not using for WorkSpaces, and try again.

registerWorkspaceDirectory_enableWorkDocs :: Lens' RegisterWorkspaceDirectory Bool

Indicates whether Amazon WorkDocs is enabled or disabled. If you have enabled this parameter and WorkDocs is not available in the Region, you will receive an OperationNotSupportedException error. Set EnableWorkDocs to disabled, and try again.

Destructuring the Response

data RegisterWorkspaceDirectoryResponse

See: newRegisterWorkspaceDirectoryResponse smart constructor.

Constructors

RegisterWorkspaceDirectoryResponse'

Fields

httpStatus :: Int

The response's http status code.

newRegisterWorkspaceDirectoryResponse

Arguments

:: Int
$sel:httpStatus:RegisterWorkspaceDirectoryResponse'

-> RegisterWorkspaceDirectoryResponse

Create a value of RegisterWorkspaceDirectoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RegisterWorkspaceDirectoryResponse', registerWorkspaceDirectoryResponse_httpStatus - The response's http status code.

Response Lenses

registerWorkspaceDirectoryResponse_httpStatus :: Lens' RegisterWorkspaceDirectoryResponse Int

The response's http status code.