Network.AWS.WorkSpaces.TerminateWorkspaces

Description

Terminates the specified WorkSpaces.

Terminating a WorkSpace is a permanent action and cannot be undone. The user's data is destroyed. If you need to archive any user data, contact AWS Support before terminating the WorkSpace.

You can terminate a WorkSpace that is in any state except SUSPENDED.

This operation is asynchronous and returns before the WorkSpaces have been completely terminated. After a WorkSpace is terminated, the TERMINATED state is returned only briefly before the WorkSpace directory metadata is cleaned up, so this state is rarely returned. To confirm that a WorkSpace is terminated, check for the WorkSpace ID by using DescribeWorkSpaces. If the WorkSpace ID isn't returned, then the WorkSpace has been successfully terminated.

Simple AD and AD Connector are made available to you free of charge to use with WorkSpaces. If there are no WorkSpaces being used with your Simple AD or AD Connector directory for 30 consecutive days, this directory will be automatically deregistered for use with Amazon WorkSpaces, and you will be charged for this directory as per the AWS Directory Services pricing terms.

To delete empty directories, see Delete the Directory for Your WorkSpaces. If you delete your Simple AD or AD Connector directory, you can always create a new one when you want to start using WorkSpaces again.

Creating a Request

data TerminateWorkspaces

See: newTerminateWorkspaces smart constructor.

Constructors

TerminateWorkspaces'

Fields

terminateWorkspaceRequests :: NonEmpty TerminateRequest

The WorkSpaces to terminate. You can specify up to 25 WorkSpaces.

newTerminateWorkspaces

Arguments

:: NonEmpty TerminateRequest
$sel:terminateWorkspaceRequests:TerminateWorkspaces'

-> TerminateWorkspaces

Create a value of TerminateWorkspaces with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:terminateWorkspaceRequests:TerminateWorkspaces', terminateWorkspaces_terminateWorkspaceRequests - The WorkSpaces to terminate. You can specify up to 25 WorkSpaces.

Request Lenses

terminateWorkspaces_terminateWorkspaceRequests :: Lens' TerminateWorkspaces (NonEmpty TerminateRequest)

The WorkSpaces to terminate. You can specify up to 25 WorkSpaces.

Destructuring the Response

data TerminateWorkspacesResponse

See: newTerminateWorkspacesResponse smart constructor.

Constructors

TerminateWorkspacesResponse'

Fields

failedRequests :: Maybe [FailedWorkspaceChangeRequest]

Information about the WorkSpaces that could not be terminated.

httpStatus :: Int

The response's http status code.

newTerminateWorkspacesResponse

Arguments

:: Int
$sel:httpStatus:TerminateWorkspacesResponse'

-> TerminateWorkspacesResponse

Create a value of TerminateWorkspacesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedRequests:TerminateWorkspacesResponse', terminateWorkspacesResponse_failedRequests - Information about the WorkSpaces that could not be terminated.

$sel:httpStatus:TerminateWorkspacesResponse', terminateWorkspacesResponse_httpStatus - The response's http status code.

Response Lenses

terminateWorkspacesResponse_failedRequests :: Lens' TerminateWorkspacesResponse (Maybe [FailedWorkspaceChangeRequest])

Information about the WorkSpaces that could not be terminated.

terminateWorkspacesResponse_httpStatus :: Lens' TerminateWorkspacesResponse Int

The response's http status code.