Network.AWS.WorkSpaces.Types.AccountModification

Description

Documentation

data AccountModification

Describes a modification to the configuration of Bring Your Own License (BYOL) for the specified account.

See: newAccountModification smart constructor.

Constructors

AccountModification'

Fields

startTime :: Maybe POSIX

The timestamp when the modification of the BYOL configuration was started.

dedicatedTenancySupport :: Maybe DedicatedTenancySupportResultEnum

The status of BYOL (whether BYOL is being enabled or disabled).

modificationState :: Maybe DedicatedTenancyModificationStateEnum

The state of the modification to the configuration of BYOL.

dedicatedTenancyManagementCidrRange :: Maybe Text

The IP address range, specified as an IPv4 CIDR block, for the management network interface used for the account.

errorCode :: Maybe Text

The error code that is returned if the configuration of BYOL cannot be modified.

errorMessage :: Maybe Text

The text of the error message that is returned if the configuration of BYOL cannot be modified.

newAccountModification :: AccountModification

Create a value of AccountModification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTime:AccountModification', accountModification_startTime - The timestamp when the modification of the BYOL configuration was started.

$sel:dedicatedTenancySupport:AccountModification', accountModification_dedicatedTenancySupport - The status of BYOL (whether BYOL is being enabled or disabled).

$sel:modificationState:AccountModification', accountModification_modificationState - The state of the modification to the configuration of BYOL.

$sel:dedicatedTenancyManagementCidrRange:AccountModification', accountModification_dedicatedTenancyManagementCidrRange - The IP address range, specified as an IPv4 CIDR block, for the management network interface used for the account.

$sel:errorCode:AccountModification', accountModification_errorCode - The error code that is returned if the configuration of BYOL cannot be modified.

$sel:errorMessage:AccountModification', accountModification_errorMessage - The text of the error message that is returned if the configuration of BYOL cannot be modified.

accountModification_startTime :: Lens' AccountModification (Maybe UTCTime)

The timestamp when the modification of the BYOL configuration was started.

accountModification_dedicatedTenancySupport :: Lens' AccountModification (Maybe DedicatedTenancySupportResultEnum)

The status of BYOL (whether BYOL is being enabled or disabled).

accountModification_modificationState :: Lens' AccountModification (Maybe DedicatedTenancyModificationStateEnum)

The state of the modification to the configuration of BYOL.

accountModification_dedicatedTenancyManagementCidrRange :: Lens' AccountModification (Maybe Text)

The IP address range, specified as an IPv4 CIDR block, for the management network interface used for the account.

accountModification_errorCode :: Lens' AccountModification (Maybe Text)

The error code that is returned if the configuration of BYOL cannot be modified.

accountModification_errorMessage :: Lens' AccountModification (Maybe Text)

The text of the error message that is returned if the configuration of BYOL cannot be modified.