Network.AWS.WorkSpaces.Types.ConnectionAlias

Description

Documentation

data ConnectionAlias

Describes a connection alias. Connection aliases are used for cross-Region redirection. For more information, see Cross-Region Redirection for Amazon WorkSpaces.

See: newConnectionAlias smart constructor.

Constructors

ConnectionAlias'

Fields

state :: Maybe ConnectionAliasState

The current state of the connection alias.

connectionString :: Maybe Text

The connection string specified for the connection alias. The connection string must be in the form of a fully qualified domain name (FQDN), such as www.example.com.

aliasId :: Maybe Text

The identifier of the connection alias.

ownerAccountId :: Maybe Text

The identifier of the AWS account that owns the connection alias.

associations :: Maybe (NonEmpty ConnectionAliasAssociation)

The association status of the connection alias.

newConnectionAlias :: ConnectionAlias

Create a value of ConnectionAlias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:ConnectionAlias', connectionAlias_state - The current state of the connection alias.

$sel:connectionString:ConnectionAlias', connectionAlias_connectionString - The connection string specified for the connection alias. The connection string must be in the form of a fully qualified domain name (FQDN), such as www.example.com.

$sel:aliasId:ConnectionAlias', connectionAlias_aliasId - The identifier of the connection alias.

$sel:ownerAccountId:ConnectionAlias', connectionAlias_ownerAccountId - The identifier of the AWS account that owns the connection alias.

$sel:associations:ConnectionAlias', connectionAlias_associations - The association status of the connection alias.

connectionAlias_state :: Lens' ConnectionAlias (Maybe ConnectionAliasState)

The current state of the connection alias.

connectionAlias_connectionString :: Lens' ConnectionAlias (Maybe Text)

The connection string specified for the connection alias. The connection string must be in the form of a fully qualified domain name (FQDN), such as www.example.com.

connectionAlias_aliasId :: Lens' ConnectionAlias (Maybe Text)

The identifier of the connection alias.

connectionAlias_ownerAccountId :: Lens' ConnectionAlias (Maybe Text)

The identifier of the AWS account that owns the connection alias.

connectionAlias_associations :: Lens' ConnectionAlias (Maybe (NonEmpty ConnectionAliasAssociation))

The association status of the connection alias.