Network.AWS.WorkSpaces.Types.Snapshot

Description

Documentation

data Snapshot

Describes a snapshot.

See: newSnapshot smart constructor.

Constructors

Snapshot'

Fields

snapshotTime :: Maybe POSIX

The time when the snapshot was created.

newSnapshot :: Snapshot

Create a value of Snapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshotTime:Snapshot', snapshot_snapshotTime - The time when the snapshot was created.

snapshot_snapshotTime :: Lens' Snapshot (Maybe UTCTime)

The time when the snapshot was created.