Network.AWS.WorkSpaces.Types.Workspace

Description

Documentation

data Workspace

Describes a WorkSpace.

See: newWorkspace smart constructor.

Constructors

Workspace'

Fields

directoryId :: Maybe Text

The identifier of the Directory Service directory for the WorkSpace.

state :: Maybe WorkspaceState

The operational state of the WorkSpace.

After a WorkSpace is terminated, the TERMINATED state is returned only briefly before the WorkSpace directory metadata is cleaned up, so this state is rarely returned. To confirm that a WorkSpace is terminated, check for the WorkSpace ID by using DescribeWorkSpaces. If the WorkSpace ID isn't returned, then the WorkSpace has been successfully terminated.

ipAddress :: Maybe Text

The IP address of the WorkSpace.

modificationStates :: Maybe [ModificationState]

The modification states of the WorkSpace.

userName :: Maybe Text

The user for the WorkSpace.

subnetId :: Maybe Text

The identifier of the subnet for the WorkSpace.

bundleId :: Maybe Text

The identifier of the bundle used to create the WorkSpace.

workspaceProperties :: Maybe WorkspaceProperties

The properties of the WorkSpace.

rootVolumeEncryptionEnabled :: Maybe Bool

Indicates whether the data stored on the root volume is encrypted.

errorCode :: Maybe Text

The error code that is returned if the WorkSpace cannot be created.

volumeEncryptionKey :: Maybe Text

The symmetric KMS key used to encrypt data stored on your WorkSpace. Amazon WorkSpaces does not support asymmetric KMS keys.

computerName :: Maybe Text

The name of the WorkSpace, as seen by the operating system. The format of this name varies. For more information, see Launch a WorkSpace.

workspaceId :: Maybe Text

The identifier of the WorkSpace.

userVolumeEncryptionEnabled :: Maybe Bool

Indicates whether the data stored on the user volume is encrypted.

errorMessage :: Maybe Text

The text of the error message that is returned if the WorkSpace cannot be created.

newWorkspace :: Workspace

Create a value of Workspace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryId:Workspace', workspace_directoryId - The identifier of the Directory Service directory for the WorkSpace.

$sel:state:Workspace', workspace_state - The operational state of the WorkSpace.

After a WorkSpace is terminated, the TERMINATED state is returned only briefly before the WorkSpace directory metadata is cleaned up, so this state is rarely returned. To confirm that a WorkSpace is terminated, check for the WorkSpace ID by using DescribeWorkSpaces. If the WorkSpace ID isn't returned, then the WorkSpace has been successfully terminated.

$sel:ipAddress:Workspace', workspace_ipAddress - The IP address of the WorkSpace.

$sel:modificationStates:Workspace', workspace_modificationStates - The modification states of the WorkSpace.

$sel:userName:Workspace', workspace_userName - The user for the WorkSpace.

$sel:subnetId:Workspace', workspace_subnetId - The identifier of the subnet for the WorkSpace.

$sel:bundleId:Workspace', workspace_bundleId - The identifier of the bundle used to create the WorkSpace.

$sel:workspaceProperties:Workspace', workspace_workspaceProperties - The properties of the WorkSpace.

$sel:rootVolumeEncryptionEnabled:Workspace', workspace_rootVolumeEncryptionEnabled - Indicates whether the data stored on the root volume is encrypted.

$sel:errorCode:Workspace', workspace_errorCode - The error code that is returned if the WorkSpace cannot be created.

$sel:volumeEncryptionKey:Workspace', workspace_volumeEncryptionKey - The symmetric KMS key used to encrypt data stored on your WorkSpace. Amazon WorkSpaces does not support asymmetric KMS keys.

$sel:computerName:Workspace', workspace_computerName - The name of the WorkSpace, as seen by the operating system. The format of this name varies. For more information, see Launch a WorkSpace.

$sel:workspaceId:Workspace', workspace_workspaceId - The identifier of the WorkSpace.

$sel:userVolumeEncryptionEnabled:Workspace', workspace_userVolumeEncryptionEnabled - Indicates whether the data stored on the user volume is encrypted.

$sel:errorMessage:Workspace', workspace_errorMessage - The text of the error message that is returned if the WorkSpace cannot be created.

workspace_directoryId :: Lens' Workspace (Maybe Text)

The identifier of the Directory Service directory for the WorkSpace.

workspace_state :: Lens' Workspace (Maybe WorkspaceState)

The operational state of the WorkSpace.

After a WorkSpace is terminated, the TERMINATED state is returned only briefly before the WorkSpace directory metadata is cleaned up, so this state is rarely returned. To confirm that a WorkSpace is terminated, check for the WorkSpace ID by using DescribeWorkSpaces. If the WorkSpace ID isn't returned, then the WorkSpace has been successfully terminated.

workspace_ipAddress :: Lens' Workspace (Maybe Text)

The IP address of the WorkSpace.

workspace_modificationStates :: Lens' Workspace (Maybe [ModificationState])

The modification states of the WorkSpace.

workspace_userName :: Lens' Workspace (Maybe Text)

The user for the WorkSpace.

workspace_subnetId :: Lens' Workspace (Maybe Text)

The identifier of the subnet for the WorkSpace.

workspace_bundleId :: Lens' Workspace (Maybe Text)

The identifier of the bundle used to create the WorkSpace.

workspace_workspaceProperties :: Lens' Workspace (Maybe WorkspaceProperties)

The properties of the WorkSpace.

workspace_rootVolumeEncryptionEnabled :: Lens' Workspace (Maybe Bool)

Indicates whether the data stored on the root volume is encrypted.

workspace_errorCode :: Lens' Workspace (Maybe Text)

The error code that is returned if the WorkSpace cannot be created.

workspace_volumeEncryptionKey :: Lens' Workspace (Maybe Text)

The symmetric KMS key used to encrypt data stored on your WorkSpace. Amazon WorkSpaces does not support asymmetric KMS keys.

workspace_computerName :: Lens' Workspace (Maybe Text)

The name of the WorkSpace, as seen by the operating system. The format of this name varies. For more information, see Launch a WorkSpace.

workspace_workspaceId :: Lens' Workspace (Maybe Text)

The identifier of the WorkSpace.

workspace_userVolumeEncryptionEnabled :: Lens' Workspace (Maybe Bool)

Indicates whether the data stored on the user volume is encrypted.

workspace_errorMessage :: Lens' Workspace (Maybe Text)

The text of the error message that is returned if the WorkSpace cannot be created.