Network.AWS.WorkSpaces.Types.WorkspaceBundle

Description

Documentation

data WorkspaceBundle

Describes a WorkSpace bundle.

See: newWorkspaceBundle smart constructor.

Constructors

WorkspaceBundle'

Fields

creationTime :: Maybe POSIX

The time when the bundle was created.

lastUpdatedTime :: Maybe POSIX

The last time that the bundle was updated.

bundleId :: Maybe Text

The identifier of the bundle.

owner :: Maybe Text

The owner of the bundle. This is the account identifier of the owner, or AMAZON if the bundle is provided by Amazon Web Services.

rootStorage :: Maybe RootStorage

The size of the root volume.

name :: Maybe Text

The name of the bundle.

imageId :: Maybe Text

The identifier of the image that was used to create the bundle.

computeType :: Maybe ComputeType

The compute type of the bundle. For more information, see Amazon WorkSpaces Bundles.

userStorage :: Maybe UserStorage

The size of the user volume.

description :: Maybe Text

The description of the bundle.

newWorkspaceBundle :: WorkspaceBundle

Create a value of WorkspaceBundle with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:WorkspaceBundle', workspaceBundle_creationTime - The time when the bundle was created.

$sel:lastUpdatedTime:WorkspaceBundle', workspaceBundle_lastUpdatedTime - The last time that the bundle was updated.

$sel:bundleId:WorkspaceBundle', workspaceBundle_bundleId - The identifier of the bundle.

$sel:owner:WorkspaceBundle', workspaceBundle_owner - The owner of the bundle. This is the account identifier of the owner, or AMAZON if the bundle is provided by Amazon Web Services.

$sel:rootStorage:WorkspaceBundle', workspaceBundle_rootStorage - The size of the root volume.

$sel:name:WorkspaceBundle', workspaceBundle_name - The name of the bundle.

$sel:imageId:WorkspaceBundle', workspaceBundle_imageId - The identifier of the image that was used to create the bundle.

$sel:computeType:WorkspaceBundle', workspaceBundle_computeType - The compute type of the bundle. For more information, see Amazon WorkSpaces Bundles.

$sel:userStorage:WorkspaceBundle', workspaceBundle_userStorage - The size of the user volume.

$sel:description:WorkspaceBundle', workspaceBundle_description - The description of the bundle.

workspaceBundle_creationTime :: Lens' WorkspaceBundle (Maybe UTCTime)

The time when the bundle was created.

workspaceBundle_lastUpdatedTime :: Lens' WorkspaceBundle (Maybe UTCTime)

The last time that the bundle was updated.

workspaceBundle_bundleId :: Lens' WorkspaceBundle (Maybe Text)

The identifier of the bundle.

workspaceBundle_owner :: Lens' WorkspaceBundle (Maybe Text)

The owner of the bundle. This is the account identifier of the owner, or AMAZON if the bundle is provided by Amazon Web Services.

workspaceBundle_rootStorage :: Lens' WorkspaceBundle (Maybe RootStorage)

The size of the root volume.

workspaceBundle_name :: Lens' WorkspaceBundle (Maybe Text)

The name of the bundle.

workspaceBundle_imageId :: Lens' WorkspaceBundle (Maybe Text)

The identifier of the image that was used to create the bundle.

workspaceBundle_computeType :: Lens' WorkspaceBundle (Maybe ComputeType)

The compute type of the bundle. For more information, see Amazon WorkSpaces Bundles.

workspaceBundle_userStorage :: Lens' WorkspaceBundle (Maybe UserStorage)

The size of the user volume.

workspaceBundle_description :: Lens' WorkspaceBundle (Maybe Text)

The description of the bundle.