Network.AWS.WorkSpaces.Types.WorkspaceCreationProperties

Description

Documentation

data WorkspaceCreationProperties

Describes the default properties that are used for creating WorkSpaces. For more information, see Update Directory Details for Your WorkSpaces.

See: newWorkspaceCreationProperties smart constructor.

Constructors

WorkspaceCreationProperties'

Fields

enableMaintenanceMode :: Maybe Bool

Indicates whether maintenance mode is enabled for your WorkSpaces. For more information, see WorkSpace Maintenance.

defaultOu :: Maybe Text

The default organizational unit (OU) for your WorkSpaces directories. This string must be the full Lightweight Directory Access Protocol (LDAP) distinguished name for the target domain and OU. It must be in the form "OU=value,DC=value,DC=value", where value is any string of characters, and the number of domain components (DCs) is two or more. For example, OU=WorkSpaces_machines,DC=machines,DC=example,DC=com.

To avoid errors, certain characters in the distinguished name must be escaped. For more information, see Distinguished Names in the Microsoft documentation.

The API doesn't validate whether the OU exists.

enableInternetAccess :: Maybe Bool

Indicates whether internet access is enabled for your WorkSpaces.

enableWorkDocs :: Maybe Bool

Indicates whether Amazon WorkDocs is enabled for your WorkSpaces.

If WorkDocs is already enabled for a WorkSpaces directory and you disable it, new WorkSpaces launched in the directory will not have WorkDocs enabled. However, WorkDocs remains enabled for any existing WorkSpaces, unless you either disable users' access to WorkDocs or you delete the WorkDocs site. To disable users' access to WorkDocs, see Disabling Users in the Amazon WorkDocs Administration Guide. To delete a WorkDocs site, see Deleting a Site in the Amazon WorkDocs Administration Guide.

If you enable WorkDocs on a directory that already has existing WorkSpaces, the existing WorkSpaces and any new WorkSpaces that are launched in the directory will have WorkDocs enabled.

customSecurityGroupId :: Maybe Text

The identifier of your custom security group.

userEnabledAsLocalAdministrator :: Maybe Bool

Indicates whether users are local administrators of their WorkSpaces.

newWorkspaceCreationProperties :: WorkspaceCreationProperties

Create a value of WorkspaceCreationProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enableMaintenanceMode:WorkspaceCreationProperties', workspaceCreationProperties_enableMaintenanceMode - Indicates whether maintenance mode is enabled for your WorkSpaces. For more information, see WorkSpace Maintenance.

$sel:defaultOu:WorkspaceCreationProperties', workspaceCreationProperties_defaultOu - The default organizational unit (OU) for your WorkSpaces directories. This string must be the full Lightweight Directory Access Protocol (LDAP) distinguished name for the target domain and OU. It must be in the form "OU=value,DC=value,DC=value", where value is any string of characters, and the number of domain components (DCs) is two or more. For example, OU=WorkSpaces_machines,DC=machines,DC=example,DC=com.

To avoid errors, certain characters in the distinguished name must be escaped. For more information, see Distinguished Names in the Microsoft documentation.

The API doesn't validate whether the OU exists.

$sel:enableInternetAccess:WorkspaceCreationProperties', workspaceCreationProperties_enableInternetAccess - Indicates whether internet access is enabled for your WorkSpaces.

$sel:enableWorkDocs:WorkspaceCreationProperties', workspaceCreationProperties_enableWorkDocs - Indicates whether Amazon WorkDocs is enabled for your WorkSpaces.

If WorkDocs is already enabled for a WorkSpaces directory and you disable it, new WorkSpaces launched in the directory will not have WorkDocs enabled. However, WorkDocs remains enabled for any existing WorkSpaces, unless you either disable users' access to WorkDocs or you delete the WorkDocs site. To disable users' access to WorkDocs, see Disabling Users in the Amazon WorkDocs Administration Guide. To delete a WorkDocs site, see Deleting a Site in the Amazon WorkDocs Administration Guide.

If you enable WorkDocs on a directory that already has existing WorkSpaces, the existing WorkSpaces and any new WorkSpaces that are launched in the directory will have WorkDocs enabled.

$sel:customSecurityGroupId:WorkspaceCreationProperties', workspaceCreationProperties_customSecurityGroupId - The identifier of your custom security group.

$sel:userEnabledAsLocalAdministrator:WorkspaceCreationProperties', workspaceCreationProperties_userEnabledAsLocalAdministrator - Indicates whether users are local administrators of their WorkSpaces.

workspaceCreationProperties_enableMaintenanceMode :: Lens' WorkspaceCreationProperties (Maybe Bool)

Indicates whether maintenance mode is enabled for your WorkSpaces. For more information, see WorkSpace Maintenance.

workspaceCreationProperties_defaultOu :: Lens' WorkspaceCreationProperties (Maybe Text)

The default organizational unit (OU) for your WorkSpaces directories. This string must be the full Lightweight Directory Access Protocol (LDAP) distinguished name for the target domain and OU. It must be in the form "OU=value,DC=value,DC=value", where value is any string of characters, and the number of domain components (DCs) is two or more. For example, OU=WorkSpaces_machines,DC=machines,DC=example,DC=com.

To avoid errors, certain characters in the distinguished name must be escaped. For more information, see Distinguished Names in the Microsoft documentation.

The API doesn't validate whether the OU exists.

workspaceCreationProperties_enableInternetAccess :: Lens' WorkspaceCreationProperties (Maybe Bool)

Indicates whether internet access is enabled for your WorkSpaces.

workspaceCreationProperties_enableWorkDocs :: Lens' WorkspaceCreationProperties (Maybe Bool)

Indicates whether Amazon WorkDocs is enabled for your WorkSpaces.

If WorkDocs is already enabled for a WorkSpaces directory and you disable it, new WorkSpaces launched in the directory will not have WorkDocs enabled. However, WorkDocs remains enabled for any existing WorkSpaces, unless you either disable users' access to WorkDocs or you delete the WorkDocs site. To disable users' access to WorkDocs, see Disabling Users in the Amazon WorkDocs Administration Guide. To delete a WorkDocs site, see Deleting a Site in the Amazon WorkDocs Administration Guide.

If you enable WorkDocs on a directory that already has existing WorkSpaces, the existing WorkSpaces and any new WorkSpaces that are launched in the directory will have WorkDocs enabled.

workspaceCreationProperties_customSecurityGroupId :: Lens' WorkspaceCreationProperties (Maybe Text)

The identifier of your custom security group.

workspaceCreationProperties_userEnabledAsLocalAdministrator :: Lens' WorkspaceCreationProperties (Maybe Bool)

Indicates whether users are local administrators of their WorkSpaces.