Network.AWS.WorkSpaces.Types.WorkspaceImage

Description

Documentation

data WorkspaceImage

Describes a WorkSpace image.

See: newWorkspaceImage smart constructor.

Constructors

WorkspaceImage'

Fields

imageId :: Maybe Text

The identifier of the image.

name :: Maybe Text

The name of the image.

state :: Maybe WorkspaceImageState

The status of the image.

description :: Maybe Text

The description of the image.

requiredTenancy :: Maybe WorkspaceImageRequiredTenancy

Specifies whether the image is running on dedicated hardware. When Bring Your Own License (BYOL) is enabled, this value is set to DEDICATED. For more information, see Bring Your Own Windows Desktop Images.

errorMessage :: Maybe Text

The text of the error message that is returned for the image.

created :: Maybe POSIX

The date when the image was created. If the image has been shared, the AWS account that the image has been shared with sees the original creation date of the image.

operatingSystem :: Maybe OperatingSystem

The operating system that the image is running.

ownerAccountId :: Maybe Text

The identifier of the AWS account that owns the image.

errorCode :: Maybe Text

The error code that is returned for the image.

newWorkspaceImage :: WorkspaceImage

Create a value of WorkspaceImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageId:WorkspaceImage', workspaceImage_imageId - The identifier of the image.

$sel:name:WorkspaceImage', workspaceImage_name - The name of the image.

$sel:state:WorkspaceImage', workspaceImage_state - The status of the image.

$sel:description:WorkspaceImage', workspaceImage_description - The description of the image.

$sel:requiredTenancy:WorkspaceImage', workspaceImage_requiredTenancy - Specifies whether the image is running on dedicated hardware. When Bring Your Own License (BYOL) is enabled, this value is set to DEDICATED. For more information, see Bring Your Own Windows Desktop Images.

$sel:errorMessage:WorkspaceImage', workspaceImage_errorMessage - The text of the error message that is returned for the image.

$sel:created:WorkspaceImage', workspaceImage_created - The date when the image was created. If the image has been shared, the AWS account that the image has been shared with sees the original creation date of the image.

$sel:operatingSystem:WorkspaceImage', workspaceImage_operatingSystem - The operating system that the image is running.

$sel:ownerAccountId:WorkspaceImage', workspaceImage_ownerAccountId - The identifier of the AWS account that owns the image.

$sel:errorCode:WorkspaceImage', workspaceImage_errorCode - The error code that is returned for the image.

workspaceImage_imageId :: Lens' WorkspaceImage (Maybe Text)

The identifier of the image.

workspaceImage_name :: Lens' WorkspaceImage (Maybe Text)

The name of the image.

workspaceImage_state :: Lens' WorkspaceImage (Maybe WorkspaceImageState)

The status of the image.

workspaceImage_description :: Lens' WorkspaceImage (Maybe Text)

The description of the image.

workspaceImage_requiredTenancy :: Lens' WorkspaceImage (Maybe WorkspaceImageRequiredTenancy)

Specifies whether the image is running on dedicated hardware. When Bring Your Own License (BYOL) is enabled, this value is set to DEDICATED. For more information, see Bring Your Own Windows Desktop Images.

workspaceImage_errorMessage :: Lens' WorkspaceImage (Maybe Text)

The text of the error message that is returned for the image.

workspaceImage_created :: Lens' WorkspaceImage (Maybe UTCTime)

The date when the image was created. If the image has been shared, the AWS account that the image has been shared with sees the original creation date of the image.

workspaceImage_operatingSystem :: Lens' WorkspaceImage (Maybe OperatingSystem)

The operating system that the image is running.

workspaceImage_ownerAccountId :: Lens' WorkspaceImage (Maybe Text)

The identifier of the AWS account that owns the image.

workspaceImage_errorCode :: Lens' WorkspaceImage (Maybe Text)

The error code that is returned for the image.