Network.AWS.WorkSpaces.Types.WorkspaceRequest

Description

Documentation

data WorkspaceRequest

Describes the information used to create a WorkSpace.

See: newWorkspaceRequest smart constructor.

Constructors

WorkspaceRequest'

Fields

workspaceProperties :: Maybe WorkspaceProperties

The WorkSpace properties.

rootVolumeEncryptionEnabled :: Maybe Bool

Indicates whether the data stored on the root volume is encrypted.

userVolumeEncryptionEnabled :: Maybe Bool

Indicates whether the data stored on the user volume is encrypted.

volumeEncryptionKey :: Maybe Text

The symmetric AWS KMS customer master key (CMK) used to encrypt data stored on your WorkSpace. Amazon WorkSpaces does not support asymmetric CMKs.

tags :: Maybe [Tag]

The tags for the WorkSpace.

directoryId :: Text

The identifier of the AWS Directory Service directory for the WorkSpace. You can use DescribeWorkspaceDirectories to list the available directories.

userName :: Text

The user name of the user for the WorkSpace. This user name must exist in the AWS Directory Service directory for the WorkSpace.

bundleId :: Text

The identifier of the bundle for the WorkSpace. You can use DescribeWorkspaceBundles to list the available bundles.

newWorkspaceRequest

Arguments

:: Text
$sel:directoryId:WorkspaceRequest'

-> Text
$sel:userName:WorkspaceRequest'

-> Text
$sel:bundleId:WorkspaceRequest'

-> WorkspaceRequest

Create a value of WorkspaceRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workspaceProperties:WorkspaceRequest', workspaceRequest_workspaceProperties - The WorkSpace properties.

$sel:rootVolumeEncryptionEnabled:WorkspaceRequest', workspaceRequest_rootVolumeEncryptionEnabled - Indicates whether the data stored on the root volume is encrypted.

$sel:userVolumeEncryptionEnabled:WorkspaceRequest', workspaceRequest_userVolumeEncryptionEnabled - Indicates whether the data stored on the user volume is encrypted.

$sel:volumeEncryptionKey:WorkspaceRequest', workspaceRequest_volumeEncryptionKey - The symmetric AWS KMS customer master key (CMK) used to encrypt data stored on your WorkSpace. Amazon WorkSpaces does not support asymmetric CMKs.

$sel:tags:WorkspaceRequest', workspaceRequest_tags - The tags for the WorkSpace.

$sel:directoryId:WorkspaceRequest', workspaceRequest_directoryId - The identifier of the AWS Directory Service directory for the WorkSpace. You can use DescribeWorkspaceDirectories to list the available directories.

$sel:userName:WorkspaceRequest', workspaceRequest_userName - The user name of the user for the WorkSpace. This user name must exist in the AWS Directory Service directory for the WorkSpace.

$sel:bundleId:WorkspaceRequest', workspaceRequest_bundleId - The identifier of the bundle for the WorkSpace. You can use DescribeWorkspaceBundles to list the available bundles.

workspaceRequest_workspaceProperties :: Lens' WorkspaceRequest (Maybe WorkspaceProperties)

The WorkSpace properties.

workspaceRequest_rootVolumeEncryptionEnabled :: Lens' WorkspaceRequest (Maybe Bool)

Indicates whether the data stored on the root volume is encrypted.

workspaceRequest_userVolumeEncryptionEnabled :: Lens' WorkspaceRequest (Maybe Bool)

Indicates whether the data stored on the user volume is encrypted.

workspaceRequest_volumeEncryptionKey :: Lens' WorkspaceRequest (Maybe Text)

The symmetric AWS KMS customer master key (CMK) used to encrypt data stored on your WorkSpace. Amazon WorkSpaces does not support asymmetric CMKs.

workspaceRequest_tags :: Lens' WorkspaceRequest (Maybe [Tag])

The tags for the WorkSpace.

workspaceRequest_directoryId :: Lens' WorkspaceRequest Text

The identifier of the AWS Directory Service directory for the WorkSpace. You can use DescribeWorkspaceDirectories to list the available directories.

workspaceRequest_userName :: Lens' WorkspaceRequest Text

The user name of the user for the WorkSpace. This user name must exist in the AWS Directory Service directory for the WorkSpace.

workspaceRequest_bundleId :: Lens' WorkspaceRequest Text

The identifier of the bundle for the WorkSpace. You can use DescribeWorkspaceBundles to list the available bundles.