Network.AWS.WorkSpaces.UpdateWorkspaceImagePermission

Description

Shares or unshares an image with one account in the same AWS Region by specifying whether that account has permission to copy the image. If the copy image permission is granted, the image is shared with that account. If the copy image permission is revoked, the image is unshared with the account.

After an image has been shared, the recipient account can copy the image to other AWS Regions as needed.

In the China (Ningxia) Region, you can copy images only within the same Region.

In the AWS GovCloud (US-West) Region, to copy images to and from other AWS Regions, contact AWS Support.

For more information about sharing images, see Share or Unshare a Custom WorkSpaces Image.

To delete an image that has been shared, you must unshare the image before you delete it.

Sharing Bring Your Own License (BYOL) images across AWS accounts isn't supported at this time in the AWS GovCloud (US-West) Region. To share BYOL images across accounts in the AWS GovCloud (US-West) Region, contact AWS Support.

Creating a Request

data UpdateWorkspaceImagePermission

See: newUpdateWorkspaceImagePermission smart constructor.

Constructors

UpdateWorkspaceImagePermission'

Fields

imageId :: Text

The identifier of the image.

allowCopyImage :: Bool

The permission to copy the image. This permission can be revoked only after an image has been shared.

sharedAccountId :: Text

The identifier of the AWS account to share or unshare the image with.

Before sharing the image, confirm that you are sharing to the correct AWS account ID.

newUpdateWorkspaceImagePermission

Arguments

:: Text
$sel:imageId:UpdateWorkspaceImagePermission'

-> Bool
$sel:allowCopyImage:UpdateWorkspaceImagePermission'

-> Text
$sel:sharedAccountId:UpdateWorkspaceImagePermission'

-> UpdateWorkspaceImagePermission

Create a value of UpdateWorkspaceImagePermission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageId:UpdateWorkspaceImagePermission', updateWorkspaceImagePermission_imageId - The identifier of the image.

$sel:allowCopyImage:UpdateWorkspaceImagePermission', updateWorkspaceImagePermission_allowCopyImage - The permission to copy the image. This permission can be revoked only after an image has been shared.

$sel:sharedAccountId:UpdateWorkspaceImagePermission', updateWorkspaceImagePermission_sharedAccountId - The identifier of the AWS account to share or unshare the image with.

Before sharing the image, confirm that you are sharing to the correct AWS account ID.

Request Lenses

updateWorkspaceImagePermission_imageId :: Lens' UpdateWorkspaceImagePermission Text

The identifier of the image.

updateWorkspaceImagePermission_allowCopyImage :: Lens' UpdateWorkspaceImagePermission Bool

The permission to copy the image. This permission can be revoked only after an image has been shared.

updateWorkspaceImagePermission_sharedAccountId :: Lens' UpdateWorkspaceImagePermission Text

The identifier of the AWS account to share or unshare the image with.

Before sharing the image, confirm that you are sharing to the correct AWS account ID.

Destructuring the Response

data UpdateWorkspaceImagePermissionResponse

See: newUpdateWorkspaceImagePermissionResponse smart constructor.

Constructors

UpdateWorkspaceImagePermissionResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateWorkspaceImagePermissionResponse

Arguments

:: Int
$sel:httpStatus:UpdateWorkspaceImagePermissionResponse'

-> UpdateWorkspaceImagePermissionResponse

Create a value of UpdateWorkspaceImagePermissionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateWorkspaceImagePermissionResponse', updateWorkspaceImagePermissionResponse_httpStatus - The response's http status code.

Response Lenses

updateWorkspaceImagePermissionResponse_httpStatus :: Lens' UpdateWorkspaceImagePermissionResponse Int

The response's http status code.