Network.AWS.XRay.GetServiceGraph

Description

Retrieves a document that describes services that process incoming requests, and downstream services that they call as a result. Root services process incoming requests and make calls to downstream services. Root services are applications that use the Amazon Web Services X-Ray SDK. Downstream services can be other applications, Amazon Web Services resources, HTTP web APIs, or SQL databases.

This operation returns paginated results.

Creating a Request

data GetServiceGraph

See: newGetServiceGraph smart constructor.

Constructors

GetServiceGraph'

Fields

nextToken :: Maybe Text

Pagination token.

groupARN :: Maybe Text

The Amazon Resource Name (ARN) of a group based on which you want to generate a graph.

groupName :: Maybe Text

The name of a group based on which you want to generate a graph.

startTime :: POSIX

The start of the time frame for which to generate a graph.

endTime :: POSIX

The end of the timeframe for which to generate a graph.

newGetServiceGraph

Arguments

:: UTCTime
$sel:startTime:GetServiceGraph'

-> UTCTime
$sel:endTime:GetServiceGraph'

-> GetServiceGraph

Create a value of GetServiceGraph with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetServiceGraph', getServiceGraph_nextToken - Pagination token.

$sel:groupARN:GetServiceGraph', getServiceGraph_groupARN - The Amazon Resource Name (ARN) of a group based on which you want to generate a graph.

$sel:groupName:GetServiceGraph', getServiceGraph_groupName - The name of a group based on which you want to generate a graph.

$sel:startTime:GetServiceGraph', getServiceGraph_startTime - The start of the time frame for which to generate a graph.

$sel:endTime:GetServiceGraph', getServiceGraph_endTime - The end of the timeframe for which to generate a graph.

Request Lenses

getServiceGraph_nextToken :: Lens' GetServiceGraph (Maybe Text)

Pagination token.

getServiceGraph_groupARN :: Lens' GetServiceGraph (Maybe Text)

The Amazon Resource Name (ARN) of a group based on which you want to generate a graph.

getServiceGraph_groupName :: Lens' GetServiceGraph (Maybe Text)

The name of a group based on which you want to generate a graph.

getServiceGraph_startTime :: Lens' GetServiceGraph UTCTime

The start of the time frame for which to generate a graph.

getServiceGraph_endTime :: Lens' GetServiceGraph UTCTime

The end of the timeframe for which to generate a graph.

Destructuring the Response

data GetServiceGraphResponse

See: newGetServiceGraphResponse smart constructor.

Constructors

GetServiceGraphResponse'

Fields

containsOldGroupVersions :: Maybe Bool

A flag indicating whether the group's filter expression has been consistent, or if the returned service graph may show traces from an older version of the group's filter expression.

startTime :: Maybe POSIX

The start of the time frame for which the graph was generated.

nextToken :: Maybe Text

Pagination token.

endTime :: Maybe POSIX

The end of the time frame for which the graph was generated.

services :: Maybe [ServiceInfo]

The services that have processed a traced request during the specified time frame.

httpStatus :: Int

The response's http status code.

newGetServiceGraphResponse

Arguments

:: Int
$sel:httpStatus:GetServiceGraphResponse'

-> GetServiceGraphResponse

Create a value of GetServiceGraphResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containsOldGroupVersions:GetServiceGraphResponse', getServiceGraphResponse_containsOldGroupVersions - A flag indicating whether the group's filter expression has been consistent, or if the returned service graph may show traces from an older version of the group's filter expression.

$sel:startTime:GetServiceGraph', getServiceGraphResponse_startTime - The start of the time frame for which the graph was generated.

$sel:nextToken:GetServiceGraph', getServiceGraphResponse_nextToken - Pagination token.

$sel:endTime:GetServiceGraph', getServiceGraphResponse_endTime - The end of the time frame for which the graph was generated.

$sel:services:GetServiceGraphResponse', getServiceGraphResponse_services - The services that have processed a traced request during the specified time frame.

$sel:httpStatus:GetServiceGraphResponse', getServiceGraphResponse_httpStatus - The response's http status code.

Response Lenses

getServiceGraphResponse_containsOldGroupVersions :: Lens' GetServiceGraphResponse (Maybe Bool)

A flag indicating whether the group's filter expression has been consistent, or if the returned service graph may show traces from an older version of the group's filter expression.

getServiceGraphResponse_startTime :: Lens' GetServiceGraphResponse (Maybe UTCTime)

The start of the time frame for which the graph was generated.

getServiceGraphResponse_nextToken :: Lens' GetServiceGraphResponse (Maybe Text)

Pagination token.

getServiceGraphResponse_endTime :: Lens' GetServiceGraphResponse (Maybe UTCTime)

The end of the time frame for which the graph was generated.

getServiceGraphResponse_services :: Lens' GetServiceGraphResponse (Maybe [ServiceInfo])

The services that have processed a traced request during the specified time frame.

getServiceGraphResponse_httpStatus :: Lens' GetServiceGraphResponse Int

The response's http status code.