Network.AWS.XRay.GetTimeSeriesServiceStatistics

Description

Get an aggregation of service statistics defined by a specific time range.

This operation returns paginated results.

Creating a Request

data GetTimeSeriesServiceStatistics

See: newGetTimeSeriesServiceStatistics smart constructor.

Constructors

GetTimeSeriesServiceStatistics'

Fields

entitySelectorExpression :: Maybe Text

A filter expression defining entities that will be aggregated for statistics. Supports ID, service, and edge functions. If no selector expression is specified, edge statistics are returned.

nextToken :: Maybe Text

Pagination token.

groupName :: Maybe Text

The case-sensitive name of the group for which to pull statistics from.

forecastStatistics :: Maybe Bool

The forecasted high and low fault count values. Forecast enabled requests require the EntitySelectorExpression ID be provided.

period :: Maybe Int

Aggregation period in seconds.

groupARN :: Maybe Text

The Amazon Resource Name (ARN) of the group for which to pull statistics from.

startTime :: POSIX

The start of the time frame for which to aggregate statistics.

endTime :: POSIX

The end of the time frame for which to aggregate statistics.

newGetTimeSeriesServiceStatistics

Arguments

:: UTCTime
$sel:startTime:GetTimeSeriesServiceStatistics'

-> UTCTime
$sel:endTime:GetTimeSeriesServiceStatistics'

-> GetTimeSeriesServiceStatistics

Create a value of GetTimeSeriesServiceStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entitySelectorExpression:GetTimeSeriesServiceStatistics', getTimeSeriesServiceStatistics_entitySelectorExpression - A filter expression defining entities that will be aggregated for statistics. Supports ID, service, and edge functions. If no selector expression is specified, edge statistics are returned.

$sel:nextToken:GetTimeSeriesServiceStatistics', getTimeSeriesServiceStatistics_nextToken - Pagination token.

$sel:groupName:GetTimeSeriesServiceStatistics', getTimeSeriesServiceStatistics_groupName - The case-sensitive name of the group for which to pull statistics from.

$sel:forecastStatistics:GetTimeSeriesServiceStatistics', getTimeSeriesServiceStatistics_forecastStatistics - The forecasted high and low fault count values. Forecast enabled requests require the EntitySelectorExpression ID be provided.

$sel:period:GetTimeSeriesServiceStatistics', getTimeSeriesServiceStatistics_period - Aggregation period in seconds.

$sel:groupARN:GetTimeSeriesServiceStatistics', getTimeSeriesServiceStatistics_groupARN - The Amazon Resource Name (ARN) of the group for which to pull statistics from.

$sel:startTime:GetTimeSeriesServiceStatistics', getTimeSeriesServiceStatistics_startTime - The start of the time frame for which to aggregate statistics.

$sel:endTime:GetTimeSeriesServiceStatistics', getTimeSeriesServiceStatistics_endTime - The end of the time frame for which to aggregate statistics.

Request Lenses

getTimeSeriesServiceStatistics_entitySelectorExpression :: Lens' GetTimeSeriesServiceStatistics (Maybe Text)

A filter expression defining entities that will be aggregated for statistics. Supports ID, service, and edge functions. If no selector expression is specified, edge statistics are returned.

getTimeSeriesServiceStatistics_nextToken :: Lens' GetTimeSeriesServiceStatistics (Maybe Text)

Pagination token.

getTimeSeriesServiceStatistics_groupName :: Lens' GetTimeSeriesServiceStatistics (Maybe Text)

The case-sensitive name of the group for which to pull statistics from.

getTimeSeriesServiceStatistics_forecastStatistics :: Lens' GetTimeSeriesServiceStatistics (Maybe Bool)

The forecasted high and low fault count values. Forecast enabled requests require the EntitySelectorExpression ID be provided.

getTimeSeriesServiceStatistics_period :: Lens' GetTimeSeriesServiceStatistics (Maybe Int)

Aggregation period in seconds.

getTimeSeriesServiceStatistics_groupARN :: Lens' GetTimeSeriesServiceStatistics (Maybe Text)

The Amazon Resource Name (ARN) of the group for which to pull statistics from.

getTimeSeriesServiceStatistics_startTime :: Lens' GetTimeSeriesServiceStatistics UTCTime

The start of the time frame for which to aggregate statistics.

getTimeSeriesServiceStatistics_endTime :: Lens' GetTimeSeriesServiceStatistics UTCTime

The end of the time frame for which to aggregate statistics.

Destructuring the Response

data GetTimeSeriesServiceStatisticsResponse

See: newGetTimeSeriesServiceStatisticsResponse smart constructor.

Constructors

GetTimeSeriesServiceStatisticsResponse'

Fields

nextToken :: Maybe Text

Pagination token.

timeSeriesServiceStatistics :: Maybe [TimeSeriesServiceStatistics]

The collection of statistics.

containsOldGroupVersions :: Maybe Bool

A flag indicating whether or not a group's filter expression has been consistent, or if a returned aggregation might show statistics from an older version of the group's filter expression.

httpStatus :: Int

The response's http status code.

newGetTimeSeriesServiceStatisticsResponse

Arguments

:: Int
$sel:httpStatus:GetTimeSeriesServiceStatisticsResponse'

-> GetTimeSeriesServiceStatisticsResponse

Create a value of GetTimeSeriesServiceStatisticsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetTimeSeriesServiceStatistics', getTimeSeriesServiceStatisticsResponse_nextToken - Pagination token.

$sel:timeSeriesServiceStatistics:GetTimeSeriesServiceStatisticsResponse', getTimeSeriesServiceStatisticsResponse_timeSeriesServiceStatistics - The collection of statistics.

$sel:containsOldGroupVersions:GetTimeSeriesServiceStatisticsResponse', getTimeSeriesServiceStatisticsResponse_containsOldGroupVersions - A flag indicating whether or not a group's filter expression has been consistent, or if a returned aggregation might show statistics from an older version of the group's filter expression.

$sel:httpStatus:GetTimeSeriesServiceStatisticsResponse', getTimeSeriesServiceStatisticsResponse_httpStatus - The response's http status code.

Response Lenses

getTimeSeriesServiceStatisticsResponse_nextToken :: Lens' GetTimeSeriesServiceStatisticsResponse (Maybe Text)

Pagination token.

getTimeSeriesServiceStatisticsResponse_timeSeriesServiceStatistics :: Lens' GetTimeSeriesServiceStatisticsResponse (Maybe [TimeSeriesServiceStatistics])

The collection of statistics.

getTimeSeriesServiceStatisticsResponse_containsOldGroupVersions :: Lens' GetTimeSeriesServiceStatisticsResponse (Maybe Bool)

A flag indicating whether or not a group's filter expression has been consistent, or if a returned aggregation might show statistics from an older version of the group's filter expression.

getTimeSeriesServiceStatisticsResponse_httpStatus :: Lens' GetTimeSeriesServiceStatisticsResponse Int

The response's http status code.