Network.AWS.XRay.PutTelemetryRecords

Description

Used by the Amazon Web Services X-Ray daemon to upload telemetry.

Creating a Request

data PutTelemetryRecords

See: newPutTelemetryRecords smart constructor.

Constructors

PutTelemetryRecords'

Fields

hostname :: Maybe Text

resourceARN :: Maybe Text

eC2InstanceId :: Maybe Text

telemetryRecords :: [TelemetryRecord]

newPutTelemetryRecords :: PutTelemetryRecords

Create a value of PutTelemetryRecords with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostname:PutTelemetryRecords', putTelemetryRecords_hostname -

$sel:resourceARN:PutTelemetryRecords', putTelemetryRecords_resourceARN -

$sel:eC2InstanceId:PutTelemetryRecords', putTelemetryRecords_eC2InstanceId -

$sel:telemetryRecords:PutTelemetryRecords', putTelemetryRecords_telemetryRecords -

Request Lenses

putTelemetryRecords_hostname :: Lens' PutTelemetryRecords (Maybe Text)

putTelemetryRecords_resourceARN :: Lens' PutTelemetryRecords (Maybe Text)

putTelemetryRecords_eC2InstanceId :: Lens' PutTelemetryRecords (Maybe Text)

putTelemetryRecords_telemetryRecords :: Lens' PutTelemetryRecords [TelemetryRecord]

Destructuring the Response

data PutTelemetryRecordsResponse

See: newPutTelemetryRecordsResponse smart constructor.

Constructors

PutTelemetryRecordsResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutTelemetryRecordsResponse

Arguments

:: Int
$sel:httpStatus:PutTelemetryRecordsResponse'

-> PutTelemetryRecordsResponse

Create a value of PutTelemetryRecordsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutTelemetryRecordsResponse', putTelemetryRecordsResponse_httpStatus - The response's http status code.

Response Lenses

putTelemetryRecordsResponse_httpStatus :: Lens' PutTelemetryRecordsResponse Int

The response's http status code.