Network.AWS.XRay.Types.AnomalousService

Description

Documentation

data AnomalousService

The service within the service graph that has anomalously high fault rates.

See: newAnomalousService smart constructor.

Constructors

AnomalousService'

Fields

serviceId :: Maybe ServiceId

newAnomalousService :: AnomalousService

Create a value of AnomalousService with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceId:AnomalousService', anomalousService_serviceId - Undocumented member.

anomalousService_serviceId :: Lens' AnomalousService (Maybe ServiceId)

Undocumented member.