Network.AWS.XRay.Types.EncryptionConfig

Description

Documentation

data EncryptionConfig

A configuration document that specifies encryption configuration settings.

See: newEncryptionConfig smart constructor.

Constructors

EncryptionConfig'

Fields

status :: Maybe EncryptionStatus

The encryption status. While the status is UPDATING, X-Ray may encrypt data with a combination of the new and old settings.

type' :: Maybe EncryptionType

The type of encryption. Set to KMS for encryption with KMS keys. Set to NONE for default encryption.

keyId :: Maybe Text

The ID of the KMS key used for encryption, if applicable.

newEncryptionConfig :: EncryptionConfig

Create a value of EncryptionConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:EncryptionConfig', encryptionConfig_status - The encryption status. While the status is UPDATING, X-Ray may encrypt data with a combination of the new and old settings.

$sel:type':EncryptionConfig', encryptionConfig_type - The type of encryption. Set to KMS for encryption with KMS keys. Set to NONE for default encryption.

$sel:keyId:EncryptionConfig', encryptionConfig_keyId - The ID of the KMS key used for encryption, if applicable.

encryptionConfig_status :: Lens' EncryptionConfig (Maybe EncryptionStatus)

The encryption status. While the status is UPDATING, X-Ray may encrypt data with a combination of the new and old settings.

encryptionConfig_type :: Lens' EncryptionConfig (Maybe EncryptionType)

The type of encryption. Set to KMS for encryption with KMS keys. Set to NONE for default encryption.

encryptionConfig_keyId :: Lens' EncryptionConfig (Maybe Text)

The ID of the KMS key used for encryption, if applicable.