Network.AWS.XRay.Types.ErrorRootCause

Description

Documentation

data ErrorRootCause

The root cause of a trace summary error.

See: newErrorRootCause smart constructor.

Constructors

ErrorRootCause'

Fields

services :: Maybe [ErrorRootCauseService]

A list of services corresponding to an error. A service identifies a segment and it contains a name, account ID, type, and inferred flag.

clientImpacting :: Maybe Bool

A flag that denotes that the root cause impacts the trace client.

newErrorRootCause :: ErrorRootCause

Create a value of ErrorRootCause with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:services:ErrorRootCause', errorRootCause_services - A list of services corresponding to an error. A service identifies a segment and it contains a name, account ID, type, and inferred flag.

$sel:clientImpacting:ErrorRootCause', errorRootCause_clientImpacting - A flag that denotes that the root cause impacts the trace client.

errorRootCause_services :: Lens' ErrorRootCause (Maybe [ErrorRootCauseService])

A list of services corresponding to an error. A service identifies a segment and it contains a name, account ID, type, and inferred flag.

errorRootCause_clientImpacting :: Lens' ErrorRootCause (Maybe Bool)

A flag that denotes that the root cause impacts the trace client.