Network.AWS.XRay.Types.ForecastStatistics

Description

Documentation

data ForecastStatistics

The predicted high and low fault count. This is used to determine if a service has become anomalous and if an insight should be created.

See: newForecastStatistics smart constructor.

Constructors

ForecastStatistics'

Fields

faultCountLow :: Maybe Integer

The lower limit of fault counts for a service.

faultCountHigh :: Maybe Integer

The upper limit of fault counts for a service.

newForecastStatistics :: ForecastStatistics

Create a value of ForecastStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:faultCountLow:ForecastStatistics', forecastStatistics_faultCountLow - The lower limit of fault counts for a service.

$sel:faultCountHigh:ForecastStatistics', forecastStatistics_faultCountHigh - The upper limit of fault counts for a service.

forecastStatistics_faultCountLow :: Lens' ForecastStatistics (Maybe Integer)

The lower limit of fault counts for a service.

forecastStatistics_faultCountHigh :: Lens' ForecastStatistics (Maybe Integer)

The upper limit of fault counts for a service.