Network.AWS.XRay.Types.Group

Description

Documentation

data Group

Details and metadata for a group.

See: newGroup smart constructor.

Constructors

Group'

Fields

groupName :: Maybe Text

The unique case-sensitive name of the group.

insightsConfiguration :: Maybe InsightsConfiguration

The structure containing configurations related to insights.

The InsightsEnabled boolean can be set to true to enable insights for the group or false to disable insights for the group.

The NotificationsEnabled boolean can be set to true to enable insights notifications through Amazon EventBridge for the group.

filterExpression :: Maybe Text

The filter expression defining the parameters to include traces.

groupARN :: Maybe Text

The Amazon Resource Name (ARN) of the group generated based on the GroupName.

newGroup :: Group

Create a value of Group with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupName:Group', group_groupName - The unique case-sensitive name of the group.

$sel:insightsConfiguration:Group', group_insightsConfiguration - The structure containing configurations related to insights.

The InsightsEnabled boolean can be set to true to enable insights for the group or false to disable insights for the group.

The NotificationsEnabled boolean can be set to true to enable insights notifications through Amazon EventBridge for the group.

$sel:filterExpression:Group', group_filterExpression - The filter expression defining the parameters to include traces.

$sel:groupARN:Group', group_groupARN - The Amazon Resource Name (ARN) of the group generated based on the GroupName.

group_groupName :: Lens' Group (Maybe Text)

The unique case-sensitive name of the group.

group_insightsConfiguration :: Lens' Group (Maybe InsightsConfiguration)

The structure containing configurations related to insights.

The InsightsEnabled boolean can be set to true to enable insights for the group or false to disable insights for the group.

The NotificationsEnabled boolean can be set to true to enable insights notifications through Amazon EventBridge for the group.

group_filterExpression :: Lens' Group (Maybe Text)

The filter expression defining the parameters to include traces.

group_groupARN :: Lens' Group (Maybe Text)

The Amazon Resource Name (ARN) of the group generated based on the GroupName.