Network.AWS.XRay.Types.Insight

Description

Documentation

data Insight

When fault rates go outside of the expected range, X-Ray creates an insight. Insights tracks emergent issues within your applications.

See: newInsight smart constructor.

Constructors

Insight'

Fields

summary :: Maybe Text

A brief description of the insight.

state :: Maybe InsightState

The current state of the insight.

startTime :: Maybe POSIX

The time, in Unix seconds, at which the insight began.

insightId :: Maybe Text

The insights unique identifier.

categories :: Maybe [InsightCategory]

The categories that label and describe the type of insight.

rootCauseServiceRequestImpactStatistics :: Maybe RequestImpactStatistics

The impact statistics of the root cause service. This includes the number of requests to the client service and whether the requests were faults or okay.

topAnomalousServices :: Maybe [AnomalousService]

The service within the insight that is most impacted by the incident.

rootCauseServiceId :: Maybe ServiceId

clientRequestImpactStatistics :: Maybe RequestImpactStatistics

The impact statistics of the client side service. This includes the number of requests to the client service and whether the requests were faults or okay.

endTime :: Maybe POSIX

The time, in Unix seconds, at which the insight ended.

groupARN :: Maybe Text

The Amazon Resource Name (ARN) of the group that the insight belongs to.

groupName :: Maybe Text

The name of the group that the insight belongs to.

newInsight :: Insight

Create a value of Insight with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:Insight', insight_summary - A brief description of the insight.

$sel:state:Insight', insight_state - The current state of the insight.

$sel:startTime:Insight', insight_startTime - The time, in Unix seconds, at which the insight began.

$sel:insightId:Insight', insight_insightId - The insights unique identifier.

$sel:categories:Insight', insight_categories - The categories that label and describe the type of insight.

$sel:rootCauseServiceRequestImpactStatistics:Insight', insight_rootCauseServiceRequestImpactStatistics - The impact statistics of the root cause service. This includes the number of requests to the client service and whether the requests were faults or okay.

$sel:topAnomalousServices:Insight', insight_topAnomalousServices - The service within the insight that is most impacted by the incident.

$sel:rootCauseServiceId:Insight', insight_rootCauseServiceId - Undocumented member.

$sel:clientRequestImpactStatistics:Insight', insight_clientRequestImpactStatistics - The impact statistics of the client side service. This includes the number of requests to the client service and whether the requests were faults or okay.

$sel:endTime:Insight', insight_endTime - The time, in Unix seconds, at which the insight ended.

$sel:groupARN:Insight', insight_groupARN - The Amazon Resource Name (ARN) of the group that the insight belongs to.

$sel:groupName:Insight', insight_groupName - The name of the group that the insight belongs to.

insight_summary :: Lens' Insight (Maybe Text)

A brief description of the insight.

insight_state :: Lens' Insight (Maybe InsightState)

The current state of the insight.

insight_startTime :: Lens' Insight (Maybe UTCTime)

The time, in Unix seconds, at which the insight began.

insight_insightId :: Lens' Insight (Maybe Text)

The insights unique identifier.

insight_categories :: Lens' Insight (Maybe [InsightCategory])

The categories that label and describe the type of insight.

insight_rootCauseServiceRequestImpactStatistics :: Lens' Insight (Maybe RequestImpactStatistics)

The impact statistics of the root cause service. This includes the number of requests to the client service and whether the requests were faults or okay.

insight_topAnomalousServices :: Lens' Insight (Maybe [AnomalousService])

The service within the insight that is most impacted by the incident.

insight_rootCauseServiceId :: Lens' Insight (Maybe ServiceId)

Undocumented member.

insight_clientRequestImpactStatistics :: Lens' Insight (Maybe RequestImpactStatistics)

The impact statistics of the client side service. This includes the number of requests to the client service and whether the requests were faults or okay.

insight_endTime :: Lens' Insight (Maybe UTCTime)

The time, in Unix seconds, at which the insight ended.

insight_groupARN :: Lens' Insight (Maybe Text)

The Amazon Resource Name (ARN) of the group that the insight belongs to.

insight_groupName :: Lens' Insight (Maybe Text)

The name of the group that the insight belongs to.