Network.AWS.XRay.Types.InsightEvent

Description

Documentation

data InsightEvent

X-Ray reevaluates insights periodically until they are resolved, and records each intermediate state in an event. You can review incident events in the Impact Timeline on the Inspect page in the X-Ray console.

See: newInsightEvent smart constructor.

Constructors

InsightEvent'

Fields

clientRequestImpactStatistics :: Maybe RequestImpactStatistics

The impact statistics of the client side service. This includes the number of requests to the client service and whether the requests were faults or okay.

rootCauseServiceRequestImpactStatistics :: Maybe RequestImpactStatistics

The impact statistics of the root cause service. This includes the number of requests to the client service and whether the requests were faults or okay.

summary :: Maybe Text

A brief description of the event.

topAnomalousServices :: Maybe [AnomalousService]

The service during the event that is most impacted by the incident.

eventTime :: Maybe POSIX

The time, in Unix seconds, at which the event was recorded.

newInsightEvent :: InsightEvent

Create a value of InsightEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientRequestImpactStatistics:InsightEvent', insightEvent_clientRequestImpactStatistics - The impact statistics of the client side service. This includes the number of requests to the client service and whether the requests were faults or okay.

$sel:rootCauseServiceRequestImpactStatistics:InsightEvent', insightEvent_rootCauseServiceRequestImpactStatistics - The impact statistics of the root cause service. This includes the number of requests to the client service and whether the requests were faults or okay.

$sel:summary:InsightEvent', insightEvent_summary - A brief description of the event.

$sel:topAnomalousServices:InsightEvent', insightEvent_topAnomalousServices - The service during the event that is most impacted by the incident.

$sel:eventTime:InsightEvent', insightEvent_eventTime - The time, in Unix seconds, at which the event was recorded.

insightEvent_clientRequestImpactStatistics :: Lens' InsightEvent (Maybe RequestImpactStatistics)

The impact statistics of the client side service. This includes the number of requests to the client service and whether the requests were faults or okay.

insightEvent_rootCauseServiceRequestImpactStatistics :: Lens' InsightEvent (Maybe RequestImpactStatistics)

The impact statistics of the root cause service. This includes the number of requests to the client service and whether the requests were faults or okay.

insightEvent_summary :: Lens' InsightEvent (Maybe Text)

A brief description of the event.

insightEvent_topAnomalousServices :: Lens' InsightEvent (Maybe [AnomalousService])

The service during the event that is most impacted by the incident.

insightEvent_eventTime :: Lens' InsightEvent (Maybe UTCTime)

The time, in Unix seconds, at which the event was recorded.