Network.AWS.XRay.Types.InsightImpactGraphService

Description

Documentation

data InsightImpactGraphService

Information about an application that processed requests, users that made requests, or downstream services, resources, and applications that an application used.

See: newInsightImpactGraphService smart constructor.

Constructors

InsightImpactGraphService'

Fields

accountId :: Maybe Text

Identifier of the Amazon Web Services account in which the service runs.

names :: Maybe [Text]

A list of names for the service, including the canonical name.

referenceId :: Maybe Int

Identifier for the service. Unique within the service map.

edges :: Maybe [InsightImpactGraphEdge]

Connections to downstream services.

name :: Maybe Text

The canonical name of the service.

type' :: Maybe Text

Identifier for the service. Unique within the service map.

Amazon Web Services Resource - The type of an Amazon Web Services resource. For example, AWS::EC2::Instance for an application running on Amazon EC2 or AWS::DynamoDB::Table for an Amazon DynamoDB table that the application used.

Amazon Web Services Service - The type of an Amazon Web Services service. For example, AWS::DynamoDB for downstream calls to Amazon DynamoDB that didn't target a specific table.

Amazon Web Services Service - The type of an Amazon Web Services service. For example, AWS::DynamoDB for downstream calls to Amazon DynamoDB that didn't target a specific table.

remote - A downstream service of indeterminate type.

newInsightImpactGraphService :: InsightImpactGraphService

Create a value of InsightImpactGraphService with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:InsightImpactGraphService', insightImpactGraphService_accountId - Identifier of the Amazon Web Services account in which the service runs.

$sel:names:InsightImpactGraphService', insightImpactGraphService_names - A list of names for the service, including the canonical name.

$sel:referenceId:InsightImpactGraphService', insightImpactGraphService_referenceId - Identifier for the service. Unique within the service map.

$sel:edges:InsightImpactGraphService', insightImpactGraphService_edges - Connections to downstream services.

$sel:name:InsightImpactGraphService', insightImpactGraphService_name - The canonical name of the service.

$sel:type':InsightImpactGraphService', insightImpactGraphService_type - Identifier for the service. Unique within the service map.

Amazon Web Services Resource - The type of an Amazon Web Services resource. For example, AWS::EC2::Instance for an application running on Amazon EC2 or AWS::DynamoDB::Table for an Amazon DynamoDB table that the application used.

Amazon Web Services Service - The type of an Amazon Web Services service. For example, AWS::DynamoDB for downstream calls to Amazon DynamoDB that didn't target a specific table.

Amazon Web Services Service - The type of an Amazon Web Services service. For example, AWS::DynamoDB for downstream calls to Amazon DynamoDB that didn't target a specific table.

remote - A downstream service of indeterminate type.

insightImpactGraphService_accountId :: Lens' InsightImpactGraphService (Maybe Text)

Identifier of the Amazon Web Services account in which the service runs.

insightImpactGraphService_names :: Lens' InsightImpactGraphService (Maybe [Text])

A list of names for the service, including the canonical name.

insightImpactGraphService_referenceId :: Lens' InsightImpactGraphService (Maybe Int)

Identifier for the service. Unique within the service map.

insightImpactGraphService_edges :: Lens' InsightImpactGraphService (Maybe [InsightImpactGraphEdge])

Connections to downstream services.

insightImpactGraphService_name :: Lens' InsightImpactGraphService (Maybe Text)

The canonical name of the service.

insightImpactGraphService_type :: Lens' InsightImpactGraphService (Maybe Text)

Identifier for the service. Unique within the service map.

Amazon Web Services Resource - The type of an Amazon Web Services resource. For example, AWS::EC2::Instance for an application running on Amazon EC2 or AWS::DynamoDB::Table for an Amazon DynamoDB table that the application used.

Amazon Web Services Service - The type of an Amazon Web Services service. For example, AWS::DynamoDB for downstream calls to Amazon DynamoDB that didn't target a specific table.

Amazon Web Services Service - The type of an Amazon Web Services service. For example, AWS::DynamoDB for downstream calls to Amazon DynamoDB that didn't target a specific table.

remote - A downstream service of indeterminate type.