Network.AWS.XRay.Types.InsightSummary

Description

Documentation

data InsightSummary

Information that describes an insight.

See: newInsightSummary smart constructor.

Constructors

InsightSummary'

Fields

clientRequestImpactStatistics :: Maybe RequestImpactStatistics

The impact statistics of the client side service. This includes the number of requests to the client service and whether the requests were faults or okay.

rootCauseServiceRequestImpactStatistics :: Maybe RequestImpactStatistics

The impact statistics of the root cause service. This includes the number of requests to the client service and whether the requests were faults or okay.

lastUpdateTime :: Maybe POSIX

The time, in Unix seconds, that the insight was last updated.

groupName :: Maybe Text

The name of the group that the insight belongs to.

startTime :: Maybe POSIX

The time, in Unix seconds, at which the insight began.

endTime :: Maybe POSIX

The time, in Unix seconds, at which the insight ended.

rootCauseServiceId :: Maybe ServiceId

state :: Maybe InsightState

The current state of the insight.

summary :: Maybe Text

A brief description of the insight.

topAnomalousServices :: Maybe [AnomalousService]

The service within the insight that is most impacted by the incident.

categories :: Maybe [InsightCategory]

Categories The categories that label and describe the type of insight.

insightId :: Maybe Text

The insights unique identifier.

groupARN :: Maybe Text

The Amazon Resource Name (ARN) of the group that the insight belongs to.

newInsightSummary :: InsightSummary

Create a value of InsightSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientRequestImpactStatistics:InsightSummary', insightSummary_clientRequestImpactStatistics - The impact statistics of the client side service. This includes the number of requests to the client service and whether the requests were faults or okay.

$sel:rootCauseServiceRequestImpactStatistics:InsightSummary', insightSummary_rootCauseServiceRequestImpactStatistics - The impact statistics of the root cause service. This includes the number of requests to the client service and whether the requests were faults or okay.

$sel:lastUpdateTime:InsightSummary', insightSummary_lastUpdateTime - The time, in Unix seconds, that the insight was last updated.

$sel:groupName:InsightSummary', insightSummary_groupName - The name of the group that the insight belongs to.

$sel:startTime:InsightSummary', insightSummary_startTime - The time, in Unix seconds, at which the insight began.

$sel:endTime:InsightSummary', insightSummary_endTime - The time, in Unix seconds, at which the insight ended.

$sel:rootCauseServiceId:InsightSummary', insightSummary_rootCauseServiceId - Undocumented member.

$sel:state:InsightSummary', insightSummary_state - The current state of the insight.

$sel:summary:InsightSummary', insightSummary_summary - A brief description of the insight.

$sel:topAnomalousServices:InsightSummary', insightSummary_topAnomalousServices - The service within the insight that is most impacted by the incident.

$sel:categories:InsightSummary', insightSummary_categories - Categories The categories that label and describe the type of insight.

$sel:insightId:InsightSummary', insightSummary_insightId - The insights unique identifier.

$sel:groupARN:InsightSummary', insightSummary_groupARN - The Amazon Resource Name (ARN) of the group that the insight belongs to.

insightSummary_clientRequestImpactStatistics :: Lens' InsightSummary (Maybe RequestImpactStatistics)

The impact statistics of the client side service. This includes the number of requests to the client service and whether the requests were faults or okay.

insightSummary_rootCauseServiceRequestImpactStatistics :: Lens' InsightSummary (Maybe RequestImpactStatistics)

The impact statistics of the root cause service. This includes the number of requests to the client service and whether the requests were faults or okay.

insightSummary_lastUpdateTime :: Lens' InsightSummary (Maybe UTCTime)

The time, in Unix seconds, that the insight was last updated.

insightSummary_groupName :: Lens' InsightSummary (Maybe Text)

The name of the group that the insight belongs to.

insightSummary_startTime :: Lens' InsightSummary (Maybe UTCTime)

The time, in Unix seconds, at which the insight began.

insightSummary_endTime :: Lens' InsightSummary (Maybe UTCTime)

The time, in Unix seconds, at which the insight ended.

insightSummary_rootCauseServiceId :: Lens' InsightSummary (Maybe ServiceId)

Undocumented member.

insightSummary_state :: Lens' InsightSummary (Maybe InsightState)

The current state of the insight.

insightSummary_summary :: Lens' InsightSummary (Maybe Text)

A brief description of the insight.

insightSummary_topAnomalousServices :: Lens' InsightSummary (Maybe [AnomalousService])

The service within the insight that is most impacted by the incident.

insightSummary_categories :: Lens' InsightSummary (Maybe [InsightCategory])

Categories The categories that label and describe the type of insight.

insightSummary_insightId :: Lens' InsightSummary (Maybe Text)

The insights unique identifier.

insightSummary_groupARN :: Lens' InsightSummary (Maybe Text)

The Amazon Resource Name (ARN) of the group that the insight belongs to.