Network.AWS.XRay.Types.ResponseTimeRootCauseService

Description

Documentation

data ResponseTimeRootCauseService

A collection of fields identifying the service in a response time warning.

See: newResponseTimeRootCauseService smart constructor.

Constructors

ResponseTimeRootCauseService'

Fields

entityPath :: Maybe [ResponseTimeRootCauseEntity]

The path of root cause entities found on the service.

accountId :: Maybe Text

The account ID associated to the service.

names :: Maybe [Text]

A collection of associated service names.

name :: Maybe Text

The service name.

inferred :: Maybe Bool

A Boolean value indicating if the service is inferred from the trace.

type' :: Maybe Text

The type associated to the service.

newResponseTimeRootCauseService :: ResponseTimeRootCauseService

Create a value of ResponseTimeRootCauseService with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entityPath:ResponseTimeRootCauseService', responseTimeRootCauseService_entityPath - The path of root cause entities found on the service.

$sel:accountId:ResponseTimeRootCauseService', responseTimeRootCauseService_accountId - The account ID associated to the service.

$sel:names:ResponseTimeRootCauseService', responseTimeRootCauseService_names - A collection of associated service names.

$sel:name:ResponseTimeRootCauseService', responseTimeRootCauseService_name - The service name.

$sel:inferred:ResponseTimeRootCauseService', responseTimeRootCauseService_inferred - A Boolean value indicating if the service is inferred from the trace.

$sel:type':ResponseTimeRootCauseService', responseTimeRootCauseService_type - The type associated to the service.

responseTimeRootCauseService_entityPath :: Lens' ResponseTimeRootCauseService (Maybe [ResponseTimeRootCauseEntity])

The path of root cause entities found on the service.

responseTimeRootCauseService_accountId :: Lens' ResponseTimeRootCauseService (Maybe Text)

The account ID associated to the service.

responseTimeRootCauseService_names :: Lens' ResponseTimeRootCauseService (Maybe [Text])

A collection of associated service names.

responseTimeRootCauseService_name :: Lens' ResponseTimeRootCauseService (Maybe Text)

The service name.

responseTimeRootCauseService_inferred :: Lens' ResponseTimeRootCauseService (Maybe Bool)

A Boolean value indicating if the service is inferred from the trace.

responseTimeRootCauseService_type :: Lens' ResponseTimeRootCauseService (Maybe Text)

The type associated to the service.