Network.AWS.XRay.Types.SamplingRule

Description

Documentation

data SamplingRule

A sampling rule that services use to decide whether to instrument a request. Rule fields can match properties of the service, or properties of a request. The service can ignore rules that don't match its properties.

See: newSamplingRule smart constructor.

Constructors

SamplingRule'

Fields

ruleName :: Maybe Text

The name of the sampling rule. Specify a rule by either name or ARN, but not both.

ruleARN :: Maybe Text

The ARN of the sampling rule. Specify a rule by either name or ARN, but not both.

attributes :: Maybe (HashMap Text Text)

Matches attributes derived from the request.

resourceARN :: Text

Matches the ARN of the Amazon Web Services resource on which the service runs.

priority :: Natural

The priority of the sampling rule.

fixedRate :: Double

The percentage of matching requests to instrument, after the reservoir is exhausted.

reservoirSize :: Natural

A fixed number of matching requests to instrument per second, prior to applying the fixed rate. The reservoir is not used directly by services, but applies to all services using the rule collectively.

serviceName :: Text

Matches the name that the service uses to identify itself in segments.

serviceType :: Text

Matches the origin that the service uses to identify its type in segments.

host :: Text

Matches the hostname from a request URL.

hTTPMethod :: Text

Matches the HTTP method of a request.

uRLPath :: Text

Matches the path from a request URL.

version :: Natural

The version of the sampling rule format (1).

newSamplingRule

Arguments

:: Text
$sel:resourceARN:SamplingRule'

-> Natural
$sel:priority:SamplingRule'

-> Double
$sel:fixedRate:SamplingRule'

-> Natural
$sel:reservoirSize:SamplingRule'

-> Text
$sel:serviceName:SamplingRule'

-> Text
$sel:serviceType:SamplingRule'

-> Text
$sel:host:SamplingRule'

-> Text
$sel:hTTPMethod:SamplingRule'

-> Text
$sel:uRLPath:SamplingRule'

-> Natural
$sel:version:SamplingRule'

-> SamplingRule

Create a value of SamplingRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleName:SamplingRule', samplingRule_ruleName - The name of the sampling rule. Specify a rule by either name or ARN, but not both.

$sel:ruleARN:SamplingRule', samplingRule_ruleARN - The ARN of the sampling rule. Specify a rule by either name or ARN, but not both.

$sel:attributes:SamplingRule', samplingRule_attributes - Matches attributes derived from the request.

$sel:resourceARN:SamplingRule', samplingRule_resourceARN - Matches the ARN of the Amazon Web Services resource on which the service runs.

$sel:priority:SamplingRule', samplingRule_priority - The priority of the sampling rule.

$sel:fixedRate:SamplingRule', samplingRule_fixedRate - The percentage of matching requests to instrument, after the reservoir is exhausted.

$sel:reservoirSize:SamplingRule', samplingRule_reservoirSize - A fixed number of matching requests to instrument per second, prior to applying the fixed rate. The reservoir is not used directly by services, but applies to all services using the rule collectively.

$sel:serviceName:SamplingRule', samplingRule_serviceName - Matches the name that the service uses to identify itself in segments.

$sel:serviceType:SamplingRule', samplingRule_serviceType - Matches the origin that the service uses to identify its type in segments.

$sel:host:SamplingRule', samplingRule_host - Matches the hostname from a request URL.

$sel:hTTPMethod:SamplingRule', samplingRule_hTTPMethod - Matches the HTTP method of a request.

$sel:uRLPath:SamplingRule', samplingRule_uRLPath - Matches the path from a request URL.

$sel:version:SamplingRule', samplingRule_version - The version of the sampling rule format (1).

samplingRule_ruleName :: Lens' SamplingRule (Maybe Text)

The name of the sampling rule. Specify a rule by either name or ARN, but not both.

samplingRule_ruleARN :: Lens' SamplingRule (Maybe Text)

The ARN of the sampling rule. Specify a rule by either name or ARN, but not both.

samplingRule_attributes :: Lens' SamplingRule (Maybe (HashMap Text Text))

Matches attributes derived from the request.

samplingRule_resourceARN :: Lens' SamplingRule Text

Matches the ARN of the Amazon Web Services resource on which the service runs.

samplingRule_priority :: Lens' SamplingRule Natural

The priority of the sampling rule.

samplingRule_fixedRate :: Lens' SamplingRule Double

The percentage of matching requests to instrument, after the reservoir is exhausted.

samplingRule_reservoirSize :: Lens' SamplingRule Natural

A fixed number of matching requests to instrument per second, prior to applying the fixed rate. The reservoir is not used directly by services, but applies to all services using the rule collectively.

samplingRule_serviceName :: Lens' SamplingRule Text

Matches the name that the service uses to identify itself in segments.

samplingRule_serviceType :: Lens' SamplingRule Text

Matches the origin that the service uses to identify its type in segments.

samplingRule_host :: Lens' SamplingRule Text

Matches the hostname from a request URL.

samplingRule_hTTPMethod :: Lens' SamplingRule Text

Matches the HTTP method of a request.

samplingRule_uRLPath :: Lens' SamplingRule Text

Matches the path from a request URL.

samplingRule_version :: Lens' SamplingRule Natural

The version of the sampling rule format (1).