Network.AWS.XRay.Types.SamplingRuleUpdate

Description

Documentation

data SamplingRuleUpdate

A document specifying changes to a sampling rule's configuration.

See: newSamplingRuleUpdate smart constructor.

Constructors

SamplingRuleUpdate'

Fields

hTTPMethod :: Maybe Text

Matches the HTTP method of a request.

resourceARN :: Maybe Text

Matches the ARN of the Amazon Web Services resource on which the service runs.

fixedRate :: Maybe Double

The percentage of matching requests to instrument, after the reservoir is exhausted.

reservoirSize :: Maybe Int

A fixed number of matching requests to instrument per second, prior to applying the fixed rate. The reservoir is not used directly by services, but applies to all services using the rule collectively.

ruleName :: Maybe Text

The name of the sampling rule. Specify a rule by either name or ARN, but not both.

ruleARN :: Maybe Text

The ARN of the sampling rule. Specify a rule by either name or ARN, but not both.

serviceName :: Maybe Text

Matches the name that the service uses to identify itself in segments.

priority :: Maybe Int

The priority of the sampling rule.

attributes :: Maybe (HashMap Text Text)

Matches attributes derived from the request.

uRLPath :: Maybe Text

Matches the path from a request URL.

host :: Maybe Text

Matches the hostname from a request URL.

serviceType :: Maybe Text

Matches the origin that the service uses to identify its type in segments.

newSamplingRuleUpdate :: SamplingRuleUpdate

Create a value of SamplingRuleUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hTTPMethod:SamplingRuleUpdate', samplingRuleUpdate_hTTPMethod - Matches the HTTP method of a request.

$sel:resourceARN:SamplingRuleUpdate', samplingRuleUpdate_resourceARN - Matches the ARN of the Amazon Web Services resource on which the service runs.

$sel:fixedRate:SamplingRuleUpdate', samplingRuleUpdate_fixedRate - The percentage of matching requests to instrument, after the reservoir is exhausted.

$sel:reservoirSize:SamplingRuleUpdate', samplingRuleUpdate_reservoirSize - A fixed number of matching requests to instrument per second, prior to applying the fixed rate. The reservoir is not used directly by services, but applies to all services using the rule collectively.

$sel:ruleName:SamplingRuleUpdate', samplingRuleUpdate_ruleName - The name of the sampling rule. Specify a rule by either name or ARN, but not both.

$sel:ruleARN:SamplingRuleUpdate', samplingRuleUpdate_ruleARN - The ARN of the sampling rule. Specify a rule by either name or ARN, but not both.

$sel:serviceName:SamplingRuleUpdate', samplingRuleUpdate_serviceName - Matches the name that the service uses to identify itself in segments.

$sel:priority:SamplingRuleUpdate', samplingRuleUpdate_priority - The priority of the sampling rule.

$sel:attributes:SamplingRuleUpdate', samplingRuleUpdate_attributes - Matches attributes derived from the request.

$sel:uRLPath:SamplingRuleUpdate', samplingRuleUpdate_uRLPath - Matches the path from a request URL.

$sel:host:SamplingRuleUpdate', samplingRuleUpdate_host - Matches the hostname from a request URL.

$sel:serviceType:SamplingRuleUpdate', samplingRuleUpdate_serviceType - Matches the origin that the service uses to identify its type in segments.

samplingRuleUpdate_hTTPMethod :: Lens' SamplingRuleUpdate (Maybe Text)

Matches the HTTP method of a request.

samplingRuleUpdate_resourceARN :: Lens' SamplingRuleUpdate (Maybe Text)

Matches the ARN of the Amazon Web Services resource on which the service runs.

samplingRuleUpdate_fixedRate :: Lens' SamplingRuleUpdate (Maybe Double)

The percentage of matching requests to instrument, after the reservoir is exhausted.

samplingRuleUpdate_reservoirSize :: Lens' SamplingRuleUpdate (Maybe Int)

A fixed number of matching requests to instrument per second, prior to applying the fixed rate. The reservoir is not used directly by services, but applies to all services using the rule collectively.

samplingRuleUpdate_ruleName :: Lens' SamplingRuleUpdate (Maybe Text)

The name of the sampling rule. Specify a rule by either name or ARN, but not both.

samplingRuleUpdate_ruleARN :: Lens' SamplingRuleUpdate (Maybe Text)

The ARN of the sampling rule. Specify a rule by either name or ARN, but not both.

samplingRuleUpdate_serviceName :: Lens' SamplingRuleUpdate (Maybe Text)

Matches the name that the service uses to identify itself in segments.

samplingRuleUpdate_priority :: Lens' SamplingRuleUpdate (Maybe Int)

The priority of the sampling rule.

samplingRuleUpdate_attributes :: Lens' SamplingRuleUpdate (Maybe (HashMap Text Text))

Matches attributes derived from the request.

samplingRuleUpdate_uRLPath :: Lens' SamplingRuleUpdate (Maybe Text)

Matches the path from a request URL.

samplingRuleUpdate_host :: Lens' SamplingRuleUpdate (Maybe Text)

Matches the hostname from a request URL.

samplingRuleUpdate_serviceType :: Lens' SamplingRuleUpdate (Maybe Text)

Matches the origin that the service uses to identify its type in segments.