Network.AWS.XRay.Types.SamplingStatisticSummary

Description

Documentation

data SamplingStatisticSummary

Aggregated request sampling data for a sampling rule across all services for a 10-second window.

See: newSamplingStatisticSummary smart constructor.

Constructors

SamplingStatisticSummary'

Fields

ruleName :: Maybe Text

The name of the sampling rule.

borrowCount :: Maybe Int

The number of requests recorded with borrowed reservoir quota.

requestCount :: Maybe Int

The number of requests that matched the rule.

sampledCount :: Maybe Int

The number of requests recorded.

timestamp :: Maybe POSIX

The start time of the reporting window.

newSamplingStatisticSummary :: SamplingStatisticSummary

Create a value of SamplingStatisticSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleName:SamplingStatisticSummary', samplingStatisticSummary_ruleName - The name of the sampling rule.

$sel:borrowCount:SamplingStatisticSummary', samplingStatisticSummary_borrowCount - The number of requests recorded with borrowed reservoir quota.

$sel:requestCount:SamplingStatisticSummary', samplingStatisticSummary_requestCount - The number of requests that matched the rule.

$sel:sampledCount:SamplingStatisticSummary', samplingStatisticSummary_sampledCount - The number of requests recorded.

$sel:timestamp:SamplingStatisticSummary', samplingStatisticSummary_timestamp - The start time of the reporting window.

samplingStatisticSummary_ruleName :: Lens' SamplingStatisticSummary (Maybe Text)

The name of the sampling rule.

samplingStatisticSummary_borrowCount :: Lens' SamplingStatisticSummary (Maybe Int)

The number of requests recorded with borrowed reservoir quota.

samplingStatisticSummary_requestCount :: Lens' SamplingStatisticSummary (Maybe Int)

The number of requests that matched the rule.

samplingStatisticSummary_sampledCount :: Lens' SamplingStatisticSummary (Maybe Int)

The number of requests recorded.

samplingStatisticSummary_timestamp :: Lens' SamplingStatisticSummary (Maybe UTCTime)

The start time of the reporting window.