Network.AWS.XRay.Types.SamplingStatisticsDocument

Description

Documentation

data SamplingStatisticsDocument

Request sampling results for a single rule from a service. Results are for the last 10 seconds unless the service has been assigned a longer reporting interval after a previous call to GetSamplingTargets.

See: newSamplingStatisticsDocument smart constructor.

Constructors

SamplingStatisticsDocument'

Fields

borrowCount :: Maybe Natural

The number of requests recorded with borrowed reservoir quota.

ruleName :: Text

The name of the sampling rule.

clientID :: Text

A unique identifier for the service in hexadecimal.

timestamp :: POSIX

The current time.

requestCount :: Natural

The number of requests that matched the rule.

sampledCount :: Natural

The number of requests recorded.

newSamplingStatisticsDocument

Arguments

:: Text
$sel:ruleName:SamplingStatisticsDocument'

-> Text
$sel:clientID:SamplingStatisticsDocument'

-> UTCTime
$sel:timestamp:SamplingStatisticsDocument'

-> Natural
$sel:requestCount:SamplingStatisticsDocument'

-> Natural
$sel:sampledCount:SamplingStatisticsDocument'

-> SamplingStatisticsDocument

Create a value of SamplingStatisticsDocument with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:borrowCount:SamplingStatisticsDocument', samplingStatisticsDocument_borrowCount - The number of requests recorded with borrowed reservoir quota.

$sel:ruleName:SamplingStatisticsDocument', samplingStatisticsDocument_ruleName - The name of the sampling rule.

$sel:clientID:SamplingStatisticsDocument', samplingStatisticsDocument_clientID - A unique identifier for the service in hexadecimal.

$sel:timestamp:SamplingStatisticsDocument', samplingStatisticsDocument_timestamp - The current time.

$sel:requestCount:SamplingStatisticsDocument', samplingStatisticsDocument_requestCount - The number of requests that matched the rule.

$sel:sampledCount:SamplingStatisticsDocument', samplingStatisticsDocument_sampledCount - The number of requests recorded.

samplingStatisticsDocument_borrowCount :: Lens' SamplingStatisticsDocument (Maybe Natural)

The number of requests recorded with borrowed reservoir quota.

samplingStatisticsDocument_ruleName :: Lens' SamplingStatisticsDocument Text

The name of the sampling rule.

samplingStatisticsDocument_clientID :: Lens' SamplingStatisticsDocument Text

A unique identifier for the service in hexadecimal.

samplingStatisticsDocument_timestamp :: Lens' SamplingStatisticsDocument UTCTime

The current time.

samplingStatisticsDocument_requestCount :: Lens' SamplingStatisticsDocument Natural

The number of requests that matched the rule.

samplingStatisticsDocument_sampledCount :: Lens' SamplingStatisticsDocument Natural

The number of requests recorded.