Network.AWS.XRay.Types.ServiceInfo

Description

Documentation

data ServiceInfo

Information about an application that processed requests, users that made requests, or downstream services, resources, and applications that an application used.

See: newServiceInfo smart constructor.

Constructors

ServiceInfo'

Fields

accountId :: Maybe Text

Identifier of the Amazon Web Services account in which the service runs.

names :: Maybe [Text]

A list of names for the service, including the canonical name.

summaryStatistics :: Maybe ServiceStatistics

Aggregated statistics for the service.

responseTimeHistogram :: Maybe [HistogramEntry]

A histogram that maps the spread of service response times.

durationHistogram :: Maybe [HistogramEntry]

A histogram that maps the spread of service durations.

referenceId :: Maybe Int

Identifier for the service. Unique within the service map.

edges :: Maybe [Edge]

Connections to downstream services.

startTime :: Maybe POSIX

The start time of the first segment that the service generated.

endTime :: Maybe POSIX

The end time of the last segment that the service generated.

name :: Maybe Text

The canonical name of the service.

state :: Maybe Text

The service's state.

root :: Maybe Bool

Indicates that the service was the first service to process a request.

type' :: Maybe Text

The type of service.

Amazon Web Services Resource - The type of an Amazon Web Services resource. For example, AWS::EC2::Instance for an application running on Amazon EC2 or AWS::DynamoDB::Table for an Amazon DynamoDB table that the application used.

Amazon Web Services Service - The type of an Amazon Web Services service. For example, AWS::DynamoDB for downstream calls to Amazon DynamoDB that didn't target a specific table.

client - Represents the clients that sent requests to a root service.

remote - A downstream service of indeterminate type.

newServiceInfo :: ServiceInfo

Create a value of ServiceInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:ServiceInfo', serviceInfo_accountId - Identifier of the Amazon Web Services account in which the service runs.

$sel:names:ServiceInfo', serviceInfo_names - A list of names for the service, including the canonical name.

$sel:summaryStatistics:ServiceInfo', serviceInfo_summaryStatistics - Aggregated statistics for the service.

$sel:responseTimeHistogram:ServiceInfo', serviceInfo_responseTimeHistogram - A histogram that maps the spread of service response times.

$sel:durationHistogram:ServiceInfo', serviceInfo_durationHistogram - A histogram that maps the spread of service durations.

$sel:referenceId:ServiceInfo', serviceInfo_referenceId - Identifier for the service. Unique within the service map.

$sel:edges:ServiceInfo', serviceInfo_edges - Connections to downstream services.

$sel:startTime:ServiceInfo', serviceInfo_startTime - The start time of the first segment that the service generated.

$sel:endTime:ServiceInfo', serviceInfo_endTime - The end time of the last segment that the service generated.

$sel:name:ServiceInfo', serviceInfo_name - The canonical name of the service.

$sel:state:ServiceInfo', serviceInfo_state - The service's state.

$sel:root:ServiceInfo', serviceInfo_root - Indicates that the service was the first service to process a request.

$sel:type':ServiceInfo', serviceInfo_type - The type of service.

Amazon Web Services Resource - The type of an Amazon Web Services resource. For example, AWS::EC2::Instance for an application running on Amazon EC2 or AWS::DynamoDB::Table for an Amazon DynamoDB table that the application used.

Amazon Web Services Service - The type of an Amazon Web Services service. For example, AWS::DynamoDB for downstream calls to Amazon DynamoDB that didn't target a specific table.

client - Represents the clients that sent requests to a root service.

remote - A downstream service of indeterminate type.

serviceInfo_accountId :: Lens' ServiceInfo (Maybe Text)

Identifier of the Amazon Web Services account in which the service runs.

serviceInfo_names :: Lens' ServiceInfo (Maybe [Text])

A list of names for the service, including the canonical name.

serviceInfo_summaryStatistics :: Lens' ServiceInfo (Maybe ServiceStatistics)

Aggregated statistics for the service.

serviceInfo_responseTimeHistogram :: Lens' ServiceInfo (Maybe [HistogramEntry])

A histogram that maps the spread of service response times.

serviceInfo_durationHistogram :: Lens' ServiceInfo (Maybe [HistogramEntry])

A histogram that maps the spread of service durations.

serviceInfo_referenceId :: Lens' ServiceInfo (Maybe Int)

Identifier for the service. Unique within the service map.

serviceInfo_edges :: Lens' ServiceInfo (Maybe [Edge])

Connections to downstream services.

serviceInfo_startTime :: Lens' ServiceInfo (Maybe UTCTime)

The start time of the first segment that the service generated.

serviceInfo_endTime :: Lens' ServiceInfo (Maybe UTCTime)

The end time of the last segment that the service generated.

serviceInfo_name :: Lens' ServiceInfo (Maybe Text)

The canonical name of the service.

serviceInfo_state :: Lens' ServiceInfo (Maybe Text)

The service's state.

serviceInfo_root :: Lens' ServiceInfo (Maybe Bool)

Indicates that the service was the first service to process a request.

serviceInfo_type :: Lens' ServiceInfo (Maybe Text)

The type of service.

Amazon Web Services Resource - The type of an Amazon Web Services resource. For example, AWS::EC2::Instance for an application running on Amazon EC2 or AWS::DynamoDB::Table for an Amazon DynamoDB table that the application used.

Amazon Web Services Service - The type of an Amazon Web Services service. For example, AWS::DynamoDB for downstream calls to Amazon DynamoDB that didn't target a specific table.

client - Represents the clients that sent requests to a root service.

remote - A downstream service of indeterminate type.