Network.AWS.XRay.Types.ServiceStatistics

Description

Documentation

data ServiceStatistics

Response statistics for a service.

See: newServiceStatistics smart constructor.

Constructors

ServiceStatistics'

Fields

faultStatistics :: Maybe FaultStatistics

Information about requests that failed with a 5xx Server Error status code.

okCount :: Maybe Integer

The number of requests that completed with a 2xx Success status code.

totalResponseTime :: Maybe Double

The aggregate response time of completed requests.

errorStatistics :: Maybe ErrorStatistics

Information about requests that failed with a 4xx Client Error status code.

totalCount :: Maybe Integer

The total number of completed requests.

newServiceStatistics :: ServiceStatistics

Create a value of ServiceStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:faultStatistics:ServiceStatistics', serviceStatistics_faultStatistics - Information about requests that failed with a 5xx Server Error status code.

$sel:okCount:ServiceStatistics', serviceStatistics_okCount - The number of requests that completed with a 2xx Success status code.

$sel:totalResponseTime:ServiceStatistics', serviceStatistics_totalResponseTime - The aggregate response time of completed requests.

$sel:errorStatistics:ServiceStatistics', serviceStatistics_errorStatistics - Information about requests that failed with a 4xx Client Error status code.

$sel:totalCount:ServiceStatistics', serviceStatistics_totalCount - The total number of completed requests.

serviceStatistics_faultStatistics :: Lens' ServiceStatistics (Maybe FaultStatistics)

Information about requests that failed with a 5xx Server Error status code.

serviceStatistics_okCount :: Lens' ServiceStatistics (Maybe Integer)

The number of requests that completed with a 2xx Success status code.

serviceStatistics_totalResponseTime :: Lens' ServiceStatistics (Maybe Double)

The aggregate response time of completed requests.

serviceStatistics_errorStatistics :: Lens' ServiceStatistics (Maybe ErrorStatistics)

Information about requests that failed with a 4xx Client Error status code.

serviceStatistics_totalCount :: Lens' ServiceStatistics (Maybe Integer)

The total number of completed requests.