Network.AWS.XRay.Types.TraceSummary

Description

Documentation

data TraceSummary

Metadata generated from the segment documents in a trace.

See: newTraceSummary smart constructor.

Constructors

TraceSummary'

Fields

instanceIds :: Maybe [InstanceIdDetail]

A list of EC2 instance IDs for any instance corresponding to the trace segments.

availabilityZones :: Maybe [AvailabilityZoneDetail]

A list of Availability Zones for any zone corresponding to the trace segments.

errorRootCauses :: Maybe [ErrorRootCause]

A collection of ErrorRootCause structures corresponding to the trace segments.

responseTime :: Maybe Double

The length of time in seconds between the start and end times of the root segment. If the service performs work asynchronously, the response time measures the time before the response is sent to the user, while the duration measures the amount of time before the last traced activity completes.

duration :: Maybe Double

The length of time in seconds between the start time of the root segment and the end time of the last segment that completed.

serviceIds :: Maybe [ServiceId]

Service IDs from the trace's segment documents.

matchedEventTime :: Maybe POSIX

The matched time stamp of a defined event.

entryPoint :: Maybe ServiceId

The root of a trace.

hasFault :: Maybe Bool

The root segment document has a 500 series error.

id :: Maybe Text

The unique identifier for the request that generated the trace's segments and subsegments.

annotations :: Maybe (HashMap Text [ValueWithServiceIds])

Annotations from the trace's segment documents.

resourceARNs :: Maybe [ResourceARNDetail]

A list of resource ARNs for any resource corresponding to the trace segments.

isPartial :: Maybe Bool

One or more of the segment documents is in progress.

faultRootCauses :: Maybe [FaultRootCause]

A collection of FaultRootCause structures corresponding to the trace segments.

revision :: Maybe Int

The revision number of a trace.

http :: Maybe Http

Information about the HTTP request served by the trace.

hasError :: Maybe Bool

The root segment document has a 400 series error.

users :: Maybe [TraceUser]

Users from the trace's segment documents.

responseTimeRootCauses :: Maybe [ResponseTimeRootCause]

A collection of ResponseTimeRootCause structures corresponding to the trace segments.

hasThrottle :: Maybe Bool

One or more of the segment documents has a 429 throttling error.

newTraceSummary :: TraceSummary

Create a value of TraceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceIds:TraceSummary', traceSummary_instanceIds - A list of EC2 instance IDs for any instance corresponding to the trace segments.

$sel:availabilityZones:TraceSummary', traceSummary_availabilityZones - A list of Availability Zones for any zone corresponding to the trace segments.

$sel:errorRootCauses:TraceSummary', traceSummary_errorRootCauses - A collection of ErrorRootCause structures corresponding to the trace segments.

$sel:responseTime:TraceSummary', traceSummary_responseTime - The length of time in seconds between the start and end times of the root segment. If the service performs work asynchronously, the response time measures the time before the response is sent to the user, while the duration measures the amount of time before the last traced activity completes.

$sel:duration:TraceSummary', traceSummary_duration - The length of time in seconds between the start time of the root segment and the end time of the last segment that completed.

$sel:serviceIds:TraceSummary', traceSummary_serviceIds - Service IDs from the trace's segment documents.

$sel:matchedEventTime:TraceSummary', traceSummary_matchedEventTime - The matched time stamp of a defined event.

$sel:entryPoint:TraceSummary', traceSummary_entryPoint - The root of a trace.

$sel:hasFault:TraceSummary', traceSummary_hasFault - The root segment document has a 500 series error.

$sel:id:TraceSummary', traceSummary_id - The unique identifier for the request that generated the trace's segments and subsegments.

$sel:annotations:TraceSummary', traceSummary_annotations - Annotations from the trace's segment documents.

$sel:resourceARNs:TraceSummary', traceSummary_resourceARNs - A list of resource ARNs for any resource corresponding to the trace segments.

$sel:isPartial:TraceSummary', traceSummary_isPartial - One or more of the segment documents is in progress.

$sel:faultRootCauses:TraceSummary', traceSummary_faultRootCauses - A collection of FaultRootCause structures corresponding to the trace segments.

$sel:revision:TraceSummary', traceSummary_revision - The revision number of a trace.

$sel:http:TraceSummary', traceSummary_http - Information about the HTTP request served by the trace.

$sel:hasError:TraceSummary', traceSummary_hasError - The root segment document has a 400 series error.

$sel:users:TraceSummary', traceSummary_users - Users from the trace's segment documents.

$sel:responseTimeRootCauses:TraceSummary', traceSummary_responseTimeRootCauses - A collection of ResponseTimeRootCause structures corresponding to the trace segments.

$sel:hasThrottle:TraceSummary', traceSummary_hasThrottle - One or more of the segment documents has a 429 throttling error.

traceSummary_instanceIds :: Lens' TraceSummary (Maybe [InstanceIdDetail])

A list of EC2 instance IDs for any instance corresponding to the trace segments.

traceSummary_availabilityZones :: Lens' TraceSummary (Maybe [AvailabilityZoneDetail])

A list of Availability Zones for any zone corresponding to the trace segments.

traceSummary_errorRootCauses :: Lens' TraceSummary (Maybe [ErrorRootCause])

A collection of ErrorRootCause structures corresponding to the trace segments.

traceSummary_responseTime :: Lens' TraceSummary (Maybe Double)

The length of time in seconds between the start and end times of the root segment. If the service performs work asynchronously, the response time measures the time before the response is sent to the user, while the duration measures the amount of time before the last traced activity completes.

traceSummary_duration :: Lens' TraceSummary (Maybe Double)

The length of time in seconds between the start time of the root segment and the end time of the last segment that completed.

traceSummary_serviceIds :: Lens' TraceSummary (Maybe [ServiceId])

Service IDs from the trace's segment documents.

traceSummary_matchedEventTime :: Lens' TraceSummary (Maybe UTCTime)

The matched time stamp of a defined event.

traceSummary_entryPoint :: Lens' TraceSummary (Maybe ServiceId)

The root of a trace.

traceSummary_hasFault :: Lens' TraceSummary (Maybe Bool)

The root segment document has a 500 series error.

traceSummary_id :: Lens' TraceSummary (Maybe Text)

The unique identifier for the request that generated the trace's segments and subsegments.

traceSummary_annotations :: Lens' TraceSummary (Maybe (HashMap Text [ValueWithServiceIds]))

Annotations from the trace's segment documents.

traceSummary_resourceARNs :: Lens' TraceSummary (Maybe [ResourceARNDetail])

A list of resource ARNs for any resource corresponding to the trace segments.

traceSummary_isPartial :: Lens' TraceSummary (Maybe Bool)

One or more of the segment documents is in progress.

traceSummary_faultRootCauses :: Lens' TraceSummary (Maybe [FaultRootCause])

A collection of FaultRootCause structures corresponding to the trace segments.

traceSummary_revision :: Lens' TraceSummary (Maybe Int)

The revision number of a trace.

traceSummary_http :: Lens' TraceSummary (Maybe Http)

Information about the HTTP request served by the trace.

traceSummary_hasError :: Lens' TraceSummary (Maybe Bool)

The root segment document has a 400 series error.

traceSummary_users :: Lens' TraceSummary (Maybe [TraceUser])

Users from the trace's segment documents.

traceSummary_responseTimeRootCauses :: Lens' TraceSummary (Maybe [ResponseTimeRootCause])

A collection of ResponseTimeRootCause structures corresponding to the trace segments.

traceSummary_hasThrottle :: Lens' TraceSummary (Maybe Bool)

One or more of the segment documents has a 429 throttling error.