Network.AWS.XRay.UpdateSamplingRule

Description

Modifies a sampling rule's configuration.

Creating a Request

data UpdateSamplingRule

See: newUpdateSamplingRule smart constructor.

Constructors

UpdateSamplingRule'

Fields

samplingRuleUpdate :: SamplingRuleUpdate

The rule and fields to change.

newUpdateSamplingRule

Arguments

:: SamplingRuleUpdate
$sel:samplingRuleUpdate:UpdateSamplingRule'

-> UpdateSamplingRule

Create a value of UpdateSamplingRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:samplingRuleUpdate:UpdateSamplingRule', updateSamplingRule_samplingRuleUpdate - The rule and fields to change.

Request Lenses

updateSamplingRule_samplingRuleUpdate :: Lens' UpdateSamplingRule SamplingRuleUpdate

The rule and fields to change.

Destructuring the Response

data UpdateSamplingRuleResponse

See: newUpdateSamplingRuleResponse smart constructor.

Constructors

UpdateSamplingRuleResponse'

Fields

samplingRuleRecord :: Maybe SamplingRuleRecord

The updated rule definition and metadata.

httpStatus :: Int

The response's http status code.

newUpdateSamplingRuleResponse

Arguments

:: Int
$sel:httpStatus:UpdateSamplingRuleResponse'

-> UpdateSamplingRuleResponse

Create a value of UpdateSamplingRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:samplingRuleRecord:UpdateSamplingRuleResponse', updateSamplingRuleResponse_samplingRuleRecord - The updated rule definition and metadata.

$sel:httpStatus:UpdateSamplingRuleResponse', updateSamplingRuleResponse_httpStatus - The response's http status code.

Response Lenses

updateSamplingRuleResponse_samplingRuleRecord :: Lens' UpdateSamplingRuleResponse (Maybe SamplingRuleRecord)

The updated rule definition and metadata.

updateSamplingRuleResponse_httpStatus :: Lens' UpdateSamplingRuleResponse Int

The response's http status code.