Control.Monad.Trans.AWS

Description

A monad transformer built on top of functions from Network.AWS which encapsulates various common parameters, errors, and usage patterns.

Transformer

type AWS = AWST IO

A convenient alias for AWST IO.

data AWST m a

The transformer. This satisfies all of the constraints that the functions in this module require, such as providing MonadResource instances, as well as keeping track of the internal Env environment.

The MonadError instance for this transformer internally uses ExceptT to handle actions that result in an Error. For more information see sendCatch and paginateCatch.

type MonadAWS m = (MonadBaseControl IO m, MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

Provides an alias for shortening type signatures if preferred.

Note: requires the ConstraintKinds extension.

Running

runAWST :: MonadBaseControl IO m => Env -> AWST m a -> m (Either Error a)

Unwrap an AWST transformer, calling all of the registered ResourceT release actions.

Environment

data Env

The environment containing the parameters required to make AWS requests.

envAuth :: Lens' Env Auth

envRegion :: Lens' Env Region

envManager :: Lens' Env Manager

envLogger :: Lens' Env Logger

scoped :: MonadReader Env m => (Env -> m a) -> m a

Pass the current environment to a function.

Creating the environment

data Credentials :: *

Constructors

FromKeys AccessKey SecretKey

FromSession AccessKey SecretKey SecurityToken

FromProfile Text

FromEnv Text Text

Discover

newEnv :: (Functor m, MonadIO m) => Region -> Credentials -> Manager -> ExceptT String m Env

This creates a new environment without debug logging and uses getAuth to expand/discover the supplied Credentials.

Lenses such as envLogger can be used to modify the Env with a debug logger.

getEnv :: Region -> Credentials -> IO Env

Create a new environment without debug logging, creating a new Manager.

Any errors are thrown using error.

See: newEnv

Debugging

debug :: (MonadIO m, MonadReader Env m) => Text -> m ()

Use the logger from envLogger to log a debug message.

whenDebug :: MonadReader Env m => m () -> m ()

Perform a monadic action if envLogger is set to Debug.

Analogous to when.

Regionalisation

data Region :: *

Constructors

Ireland

Tokyo

Singapore

Sydney

Beijing

NorthVirginia

NorthCalifornia

Oregon

GovCloud

GovCloudFIPS

SaoPaulo

within :: MonadReader Env m => Region -> m a -> m a

Scope a monadic action within the specific Region.

Errors

type Error = ServiceError String

The top-level error type.

hoistEither :: (MonadError Error m, AWSError e) => Either e a -> m a

Hoist an Either throwing the Left case, and returning the Right.

throwAWSError :: (MonadError Error m, AWSError e) => e -> m a

verify :: (AWSError e, MonadError Error m) => Prism' e a -> e -> m ()

verifyWith :: (AWSError e, MonadError Error m) => Prism' e a -> (a -> Bool) -> e -> m ()

Requests

Synchronous

send :: (MonadCatch m, MonadResource m, MonadReader Env m, MonadError Error m, AWSRequest a) => a -> m (Rs a)

Send a data type which is an instance of AWSRequest, returning it's associated Rs response type.

This will throw any HTTPException or AWSServiceError returned by the service using the MonadError instance. In the case of AWST this will cause the internal ExceptT to short-circuit and return an Error in the Left case as the result of the computation.

send_ :: (MonadCatch m, MonadResource m, MonadReader Env m, MonadError Error m, AWSRequest a) => a -> m ()

A variant of send which discards any successful response.

sendCatch :: (MonadCatch m, MonadResource m, MonadReader Env m, AWSRequest a) => a -> m (Response a)

Send a data type which is an instance of AWSRequest, returning either the associated Rs response type in the success case, or the related service's Er type in the error case.

This includes HTTPExceptions, serialisation errors, and any service errors returned as part of the Response.

Paginated

paginate :: (MonadCatch m, MonadResource m, MonadReader Env m, MonadError Error m, AWSPager a) => a -> Source m (Rs a)

Send a data type which is an instance of AWSPager and paginate while there are more results as defined by the related service operation.

Errors will be handle identically to send.

Note: The ResumableSource will close when there are no more results or the ResourceT computation is unwrapped. See: runResourceT for more information.

paginateCatch :: (MonadCatch m, MonadResource m, MonadReader Env m, AWSPager a) => a -> Source m (Response a)

Send a data type which is an instance of AWSPager and paginate over the associated Rs response type in the success case, or the related service's Er type in the error case.

Note: The ResumableSource will close when there are no more results or the ResourceT computation is unwrapped. See: runResourceT for more information.

Pre-signing URLs

presign

Arguments

:: (MonadIO m, MonadReader Env m, AWSRequest a, AWSPresigner (Sg (Sv a)))

=> a
Request to presign.

-> UTCTime
Signing time.

-> Int
Expiry time in seconds.

-> m (Signed a (Sg (Sv a)))

Presign a URL with expiry to be used at a later time.

Note: Requires the service's signer to be an instance of AWSPresigner. Not all signing process support this.

Types