Network.AWS

Description

This module provides simple Env and IO-based operations which can be performed against remote Amazon Web Services APIs, for use with the types supplied by the various amazonka-* libraries.

Usage

The key functions dealing with the request/response lifecycle are:

send, sendThrow

paginate, paginateThrow

await, awaitThrow

These functions have constraints that types from the amazonka-* libraries satisfy. To utilise these, you will need to specify what Region you wish to operate in and your Amazon credentials for AuthN/AuthZ purposes.

Credentials can be supplied in a number of ways. Either via explicit keys, via session profiles, or have Amazonka retrieve the credentials from an underlying IAM Role/Profile.

As a basic example, you might wish to store an object in an S3 bucket using amazonka-s3:

{-# LANGUAGE OverloadedStrings #-}

import qualified Network.AWS as AWS
import qualified Network.AWS.S3 as S3
import qualified System.IO as IO

example :: IO S3.PutObjectResponse
example = do
    -- A new Logger to replace the default noop logger is created, with the logger
    -- set to print debug information and errors to stdout:
    logger <- AWS.newLogger AWS.Debug IO.stdout

    -- To specify configuration preferences, newEnv is used to create a new
    -- configuration environment. The Credentials parameter is used to specify
    -- mechanism for supplying or retrieving AuthN/AuthZ information.
    -- In this case Discover will cause the library to try a number of options such
    -- as default environment variables, or an instance's IAM Profile and identity document:
    discover <- AWS.newEnv AWS.Discover

    let env =
            discover
                { AWS.envLogger = logger
                , AWS.envRegion = AWS.Frankfurt
                }

    -- The payload (and hash) for the S3 object is retrieved from a FilePath,
    -- either hashedFile or chunkedFile can be used, with the latter ensuring
    -- the contents of the file is enumerated exactly once, during send:
    body <- AWS.chunkedFile AWS.defaultChunkSize "local/path/to/object-payload"

    -- We now run the AWS computation with the overriden logger, performing the
    -- PutObject request. envRegion or within can be used to set the
    -- remote AWS Region:
    AWS.runResourceT $
        AWS.send (S3.putObject "bucket-name" "object-key" body)

Authentication and Environment

data Env

The environment containing the parameters required to make AWS requests.

Constructors

Env

Fields

envRegion :: Region

envLogger :: Logger

envRetryCheck :: Int -> HttpException -> Bool

envOverride :: Dual (Endo Service)

envManager :: Manager

envAuth :: Auth

newEnv

Arguments

:: MonadIO m

=> Credentials
Credential discovery mechanism.

-> m Env

Creates a new environment with a new Manager without debug logging and uses getAuth to expand/discover the supplied Credentials. Lenses from AWSEnv can be used to further configure the resulting Env.

Since: 1.5.0 - The region is now retrieved from the AWS_REGION environment variable (identical to official SDKs), or defaults to us-east-1. You can override the Env region by using $sel:envRegion:Env, or the current operation's region by using within.

Since: 1.3.6 - The default logic for retrying HttpExceptions now uses retryConnectionFailure to retry specific connection failure conditions up to 3 times. Previously only service specific errors were automatically retried. This can be reverted to the old behaviour by resetting the Env using $sel:envRetryCheck:Env lens to (\_ _ -> False).

Throws AuthError when environment variables or IAM profiles cannot be read.

See: newEnvWith.

newEnvWith

Arguments

:: MonadIO m

=> Credentials
Credential discovery mechanism.

-> Manager
Preload the EC2 instance check.

-> m Env

See: newEnv

The Maybe Bool parameter is used by the EC2 instance check. By passing a value of Nothing, the check will be performed. Just True would cause the check to be skipped and the host treated as an EC2 instance.

Throws AuthError when environment variables or IAM profiles cannot be read.

Service Configuration

When a request is sent, various values such as the endpoint, retry strategy, timeout and error handlers are taken from the associated Service for a request. For example, DynamoDB will use the defaultService configuration when sending PutItem, Query and all other operations.

You can modify a specific Service's default configuration by using configure or reconfigure. To modify all configurations simultaneously, see override.

An example of how you might alter default configuration using these mechanisms is demonstrated below. Firstly, the default dynamoDB service is configured to use non-SSL localhost as the endpoint:

import qualified Network.AWS as AWS
import qualified Network.AWS.DynamoDB as Dynamo

let dynamo :: AWS.Service
    dynamo = AWS.setEndpoint False "localhost" 8000 DynamoDB.defaultService

The updated configuration is then passed to the Env during setup:

env <- AWS.configure dynamo <$> AWS.newEnv AWS.Discover

AWS.runResourceT $ do
    -- This S3 operation will communicate with remote AWS APIs.
    x <- AWS.send env listBuckets

    -- DynamoDB operations will communicate with localhost:8000.
    y <- AWS.send env Dynamo.listTables

    -- Any operations for services other than DynamoDB, are not affected.
    ...

You can also scope the service configuration modifications to specific actions:

env <- AWS.newEnv AWS.Discover

AWS.runResourceT $ do
    -- Service operations here will communicate with AWS, even remote DynamoDB.
    x <- AWS.send env Dynamo.listTables

    -- Here DynamoDB operations will communicate with localhost:8000.
    y <- AWS.send (AWS.configure dynamo) Dynamo.listTables

Functions such as within, once, and timeout can also be used to modify service configuration for all (or specific) requests.

override :: (Service -> Service) -> Env -> Env

Provide a function which will be added to the existing stack of overrides applied to all service configurations.

configure :: Service -> Env -> Env

Configure a specific service. All requests belonging to the supplied service will use this configuration instead of the default.

It's suggested you modify the default service configuration, such as Network.AWS.DynamoDB.dynamoDB.

within :: Region -> Env -> Env

Scope an action within the specific Region.

once :: Env -> Env

Scope an action such that any retry logic for the Service is ignored and any requests will at most be sent once.

timeout :: Seconds -> Env -> Env

Scope an action such that any HTTP response will use this timeout value.

Default timeouts are chosen by considering:

This timeout, if set.

The related Service timeout for the sent request if set. (Usually 70s)

The $sel:envManager:Env timeout if set.

The default ClientRequest timeout. (Approximately 30s)

Running AWS Actions

runResourceT :: MonadUnliftIO m => ResourceT m a -> m a

Unwrap a ResourceT transformer, and call all registered release actions.

Note that there is some reference counting involved due to resourceForkIO. If multiple threads are sharing the same collection of resources, only the last call to runResourceT will deallocate the resources.

NOTE Since version 1.2.0, this function will throw a ResourceCleanupException if any of the cleanup functions throw an exception.

Since: resourcet-0.3.0

Credential Discovery

newtype AccessKey

An access key ID.

For example: AKIAIOSFODNN7EXAMPLE

See: Understanding and Getting Your Security Credentials.

Constructors

AccessKey ByteString

newtype SecretKey

Secret access key credential.

For example: wJalrXUtnFEMIK7MDENGbPxRfiCYEXAMPLEKE

See: Understanding and Getting Your Security Credentials.

Constructors

SecretKey ByteString

newtype SessionToken

A session token used by STS to temporarily authorise access to an AWS resource.

See: Temporary Security Credentials.

Constructors

SessionToken ByteString

data Credentials

Determines how AuthN/AuthZ information is retrieved.

Constructors

FromKeys AccessKey SecretKey
Explicit access and secret keys. See fromKeys.

FromSession AccessKey SecretKey SessionToken
Explicit access key, secret key and a session token. See fromSession.

FromEnv Text Text (Maybe Text) (Maybe Text)
Lookup specific environment variables for access key, secret key, an optional session token, and an optional region, respectively.

FromProfile Text
An IAM Profile name to lookup from the local EC2 instance-data. Environment variables to lookup for the access key, secret key and optional session token.

FromFile Text FilePath FilePath
A credentials profile name (the INI section), the path to the AWS credentials file, and the path to the ~.awsconfig file.

FromContainer
Obtain credentials by attempting to contact the ECS container agent at http://169.254.170.2 using the path in envContainerCredentialsURI. See IAM Roles for Tasks in the AWS documentation for more information.

Discover
Attempt credentials discovery via the following steps:

Read the envAccessKey, envSecretKey, and envRegion from the environment if they are set.

Read the credentials file if credFile exists.

Obtain credentials from the ECS container agent if envContainerCredentialsURI is set.

Retrieve the first available IAM profile and read the Region from the instance identity document, if running on EC2.

An attempt is made to resolve http://instance-data rather than directly retrieving http://169.254.169.254 for IAM profile information. This assists in ensuring the DNS lookup terminates promptly if not running on EC2.

AuthN/AuthZ information is handled similarly to other AWS SDKs. You can read some of the options available here.

When running on an EC2 instance and using FromProfile or Discover, a thread is forked which transparently handles the expiry and subsequent refresh of IAM profile information. See fromProfileName for more information.

Supported Regions

newtype Region

The available AWS regions.

Constructors

Region'

Fields

fromRegion :: Text

Bundled Patterns

pattern NorthVirginia :: Region

pattern Ohio :: Region

pattern NorthCalifornia :: Region

pattern Oregon :: Region

pattern GovCloudWest :: Region

pattern GovCloudEast :: Region

pattern Montreal :: Region

pattern SaoPaulo :: Region

pattern Frankfurt :: Region

pattern Ireland :: Region

pattern London :: Region

pattern Milan :: Region

pattern Paris :: Region

pattern Stockholm :: Region

pattern Bahrain :: Region

pattern CapeTown :: Region

pattern Beijing :: Region

pattern Ningxia :: Region

pattern HongKong :: Region

pattern Tokyo :: Region

pattern Seoul :: Region

pattern Osaka :: Region

pattern Singapore :: Region

pattern Sydney :: Region

pattern Mumbai :: Region

Service Endpoints

data Endpoint

Constructors

Endpoint

Fields

_endpointHost :: ByteString

_endpointSecure :: Bool

_endpointPort :: Int

_endpointScope :: ByteString

setEndpoint

Arguments

:: Bool
Whether to use HTTPS (ie. SSL).

-> ByteString
The hostname to connect to.

-> Int
The port number to connect to.

-> Service
The service configuration to override.

-> Service

A convenience function for overriding the Service Endpoint.

See: serviceEndpoint.

Sending Requests

To send a request you need to create a value of the desired operation type using the relevant constructor, as well as any further modifications of default/optional parameters using the appropriate lenses. This value can then be sent using send or paginate and the library will take care of serialisation/authentication and so forth.

The default Service configuration for a request contains retry configuration that is used to determine if a request can safely be retried and what kind of back off/on strategy should be used. (Usually exponential.) Typically services define retry strategies that handle throttling, general server errors and transport errors. Streaming requests are never retried.

send :: (MonadResource m, AWSRequest a) => Env -> a -> m (AWSResponse a)

Send a request, returning the associated response if successful.

Errors are thrown in IO.

See sendEither.

sendEither :: (MonadResource m, AWSRequest a) => Env -> a -> m (Either Error (AWSResponse a))

Send a request, returning the associated response if successful.

See send.

Pagination

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

paginate :: (MonadResource m, AWSPager a) => Env -> a -> ConduitM () (AWSResponse a) m ()

Repeatedly send a request, automatically setting markers and performing pagination. Exits on the first encountered error.

Errors are thrown in IO.

See paginateEither.

paginateEither :: (MonadResource m, AWSPager a) => Env -> a -> ConduitM () (AWSResponse a) m (Either Error ())

Repeatedly send a request, automatically setting markers and performing pagination.

Exits on the first encountered error.

See paginate.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies. Error conditions that are not handled by the Wait configuration will be thrown, or the first successful response that fulfills the success condition will be returned.

Wait specifications can be found under the Network.AWS.{ServiceName}.Waiters namespace for services which support await.

await :: (MonadResource m, AWSRequest a) => Env -> Wait a -> a -> m Accept

Poll the API with the supplied request until a specific Wait condition is fulfilled.

Errors are thrown in IO.

See awaitEither.

awaitEither :: (MonadResource m, AWSRequest a) => Env -> Wait a -> a -> m (Either Error Accept)

Poll the API with the supplied request until a specific Wait condition is fulfilled.

See await.

Streaming

Streaming comes in two flavours. HashedBody represents a request that requires a precomputed SHA256 hash, or a ChunkedBody type for those services that can perform incremental signing and do not require the entire payload to be hashed (such as S3). The type signatures for request smart constructors advertise which respective body type is required, denoting the underlying signing capabilities.

ToHashedBody and ToBody typeclass instances are available to construct the streaming bodies, automatically calculating any hash or size as needed for types such as Text, ByteString, or Aeson's Value type. To read files and other IO primitives, functions such as hashedFile, chunkedFile, or hashedBody should be used.

For responses that contain streaming bodies (such as GetObject), you can use sinkBody to connect the response body to a conduit compatible sink.

class ToBody a where

Anything that can be converted to a streaming request Body.

Minimal complete definition

Nothing

Methods

toBody :: a -> RequestBody

Convert a value to a request body.

default toBody :: ToHashedBody a => a -> RequestBody

data RequestBody

Invariant: only services that support _both_ standard and chunked signing expose RequestBody as a parameter.

Constructors

Chunked ChunkedBody

Hashed HashedBody

newtype ResponseBody

A streaming, exception safe response body.

Constructors

ResponseBody

Fields

_streamBody :: ConduitM () ByteString (ResourceT IO) ()

Hashed Request Bodies

class ToHashedBody a where

Anything that can be safely converted to a HashedBody.

Methods

toHashed :: a -> HashedBody

Convert a value to a hashed request body.

data HashedBody

An opaque request body containing a SHA256 hash.

Constructors

HashedStream (Digest SHA256) !Integer (ConduitM () ByteString (ResourceT IO) ())

HashedBytes (Digest SHA256) ByteString

hashedFile

Arguments

:: MonadIO m

=> FilePath
The file path to read.

-> m HashedBody

Construct a HashedBody from a FilePath, calculating the SHA256 hash and file size.

Note: While this function will perform in constant space, it will enumerate the entirety of the file contents _twice_. Firstly to calculate the SHA256 and lastly to stream the contents to the socket during sending.

See: ToHashedBody.

hashedFileRange

Arguments

:: MonadIO m

=> FilePath
The file path to read.

-> Integer
The byte offset at which to start reading.

-> Integer
The maximum number of bytes to read.

-> m HashedBody

Construct a HashedBody from a FilePath, specifying the range of bytes to read. This can be useful for constructing multiple requests from a single file, say for S3 multipart uploads.

See: hashedFile, sourceFileRange.

hashedBody

Arguments

:: Digest SHA256
A SHA256 hash of the file contents.

-> Integer
The size of the stream in bytes.

-> ConduitM () ByteString (ResourceT IO) ()

-> HashedBody

Construct a HashedBody from a Source, manually specifying the SHA256 hash and file size. It's left up to the caller to calculate these correctly, otherwise AWS will return signing errors.

See: ToHashedBody.

Chunked Request Bodies

data ChunkedBody

An opaque request body which will be transmitted via Transfer-Encoding: chunked.

Invariant: Only services that support chunked encoding can accept a ChunkedBody. (Currently S3.) This is enforced by the type signatures emitted by the generator.

Constructors

ChunkedBody

Fields

_chunkedSize :: ChunkSize

_chunkedLength :: Integer

_chunkedBody :: ConduitM () ByteString (ResourceT IO) ()

newtype ChunkSize

Specifies the transmitted size of the 'Transfer-Encoding' chunks.

See: defaultChunk.

Constructors

ChunkSize Int

defaultChunkSize :: ChunkSize

The default chunk size of 128 KB. The minimum chunk size accepted by AWS is 8 KB, unless the entirety of the request is below this threshold.

A chunk size of 64 KB or higher is recommended for performance reasons.

chunkedFile :: MonadIO m => ChunkSize -> FilePath -> m RequestBody

Construct a ChunkedBody from a FilePath, where the contents will be read and signed incrementally in chunks if the target service supports it.

Will intelligently revert to HashedBody if the file is smaller than the specified ChunkSize.

See: ToBody.

chunkedFileRange

Arguments

:: MonadIO m

=> ChunkSize
The idealized size of chunks that will be yielded downstream.

-> FilePath
The file path to read.

-> Integer
The byte offset at which to start reading.

-> Integer
The maximum number of bytes to read.

-> m RequestBody

Construct a ChunkedBody from a FilePath, specifying the range of bytes to read. This can be useful for constructing multiple requests from a single file, say for S3 multipart uploads.

See: chunkedFile.

unsafeChunkedBody

Arguments

:: ChunkSize
The idealized size of chunks that will be yielded downstream.

-> Integer
The size of the stream in bytes.

-> ConduitM () ByteString (ResourceT IO) ()

-> RequestBody

Unsafely construct a ChunkedBody.

This function is marked unsafe because it does nothing to enforce the chunk size. Typically for conduit IO functions, it's whatever ByteString's defaultBufferSize is, around 32 KB. If the chunk size is less than 8 KB, the request will error. 64 KB or higher chunk size is recommended for performance reasons.

Note that it will always create a chunked body even if the request is too small.

See: ToBody.

Response Bodies

sinkBody :: MonadIO m => ResponseBody -> ConduitM ByteString Void (ResourceT IO) a -> m a

Connect a Sink to a response stream.

File Size and MD5/SHA256

getFileSize :: MonadIO m => FilePath -> m Integer

Convenience function for obtaining the size of a file.

sinkMD5 :: Monad m => ConduitM ByteString o m (Digest MD5)

Incrementally calculate a MD5 Digest.

sinkSHA256 :: Monad m => ConduitM ByteString o m (Digest SHA256)

Incrementally calculate a SHA256 Digest.

Presigning Requests

Presigning requires the Service signer to be an instance of AWSPresigner. Not all signing algorithms support this.

presignURL

Arguments

:: (MonadIO m, AWSRequest a)

=> Env

-> UTCTime
Signing time.

-> Seconds
Expiry time.

-> a
Request to presign.

-> m ByteString

Presign an URL that is valid from the specified time until the number of seconds expiry has elapsed.

presign

Arguments

:: (MonadIO m, AWSRequest a)

=> Env

-> UTCTime
Signing time.

-> Seconds
Expiry time.

-> a
Request to presign.

-> m ClientRequest

Presign an HTTP request that is valid from the specified time until the number of seconds expiry has elapsed.

EC2 Instance Metadata

Metadata can be retrieved from the underlying host assuming that you're running the code on an EC2 instance or have a compatible instance-data endpoint available.

data Dynamic

Constructors

FWS
Value showing whether the customer has enabled detailed one-minute monitoring in CloudWatch.

Valid values: enabled | disabled.

Document
JSON containing instance attributes, such as instance-id, private IP address, etc. See: identity, InstanceDocument.

PKCS7
Used to verify the document's authenticity and content against the signature.

Signature

dynamic :: MonadIO m => Env -> Dynamic -> m ByteString

Retrieve the specified Dynamic data.

data Metadata

Constructors

AMIId
The AMI ID used to launch the instance.

AMILaunchIndex
If you started more than one instance at the same time, this value indicates the order in which the instance was launched. The value of the first instance launched is 0.

AMIManifestPath
The path to the AMI's manifest file in Amazon S3. If you used an Amazon EBS-backed AMI to launch the instance, the returned result is unknown.

AncestorAMIIds
The AMI IDs of any instances that were rebundled to create this AMI. This value will only exist if the AMI manifest file contained an ancestor-amis key.

BlockDevice !Mapping
See: Mapping

Hostname
The private hostname of the instance. In cases where multiple network interfaces are present, this refers to the eth0 device (the device for which the device number is 0).

IAM !Info
See: Info

InstanceAction
Notifies the instance that it should reboot in preparation for bundling. Valid values: none | shutdown | bundle-pending.

InstanceId
The ID of this instance.

InstanceType
The type of instance.

See: InstanceType

KernelId
The ID of the kernel launched with this instance, if applicable.

LocalHostname
The private DNS hostname of the instance. In cases where multiple network interfaces are present, this refers to the eth0 device (the device for which the device number is 0).

LocalIPV4
The private IP address of the instance. In cases where multiple network interfaces are present, this refers to the eth0 device (the device for which the device number is 0).

MAC
The instance's media access control (MAC) address. In cases where multiple network interfaces are present, this refers to the eth0 device (the device for which the device number is 0).

Network !Text !Interface
See: Interface

AvailabilityZone
The Availability Zone in which the instance launched.

ProductCodes
Product codes associated with the instance, if any.

PublicHostname
The instance's public DNS. If the instance is in a VPC, this category is only returned if the enableDnsHostnames attribute is set to true. For more information, see Using DNS with Your VPC.

PublicIPV4
The public IP address. If an Elastic IP address is associated with the instance, the value returned is the Elastic IP address.

OpenSSHKey
Public key. Only available if supplied at instance launch time.

RAMDiskId
The ID of the RAM disk specified at launch time, if applicable.

ReservationId
ID of the reservation.

SecurityGroups
The names of the security groups applied to the instance.

metadata :: MonadIO m => Env -> Metadata -> m ByteString

Retrieve the specified Metadata.

userdata :: MonadIO m => Env -> m (Maybe ByteString)

Retrieve the user data. Returns Nothing if no user data is assigned to the instance.

Running Asynchronous Actions

Requests can be sent asynchronously, but due to guarantees about resource closure require the use of lifted-async.

The following example demonstrates retrieving two objects from S3 concurrently:

{-# LANGUAGE OverloadedStrings #-}
import qualified Control.Concurrent.Async as Async
import qualfiied Network.AWS as AWS
import qualfiied Network.AWS.S3 as S3

let a = S3.getObject "bucket" "prefix/object-a"
let b = S3.getObject "bucket" "prefix/object-b"

Async.withAsync (AWS.runResourceT (AWS.send env a)) $ \aAsync ->
    Async.withAsync (AWS.runResourceT (AWS.send env b)) $ \bAsync -> do
        a' <- Async.wait aAsync
        b' <- Async.wait bAsync
        ...

See Control.Concurrent.Async

Handling Errors

Errors are either returned or thrown by the library using IO. Sub-errors of the canonical Error type can be caught using trying or catching and the appropriate AsError Prism when using the non-Either send variants:

trying _Error          (send $ ListObjects "bucket-name") :: Either Error          ListObjectsResponse
trying _TransportError (send $ ListObjects "bucket-name") :: Either HttpException  ListObjectsResponse
trying _SerializeError (send $ ListObjects "bucket-name") :: Either SerializeError ListObjectsResponse
trying _ServiceError   (send $ ListObjects "bucket-name") :: Either ServiceError   ListObjectsResponse

Many of the individual amazonka-* libraries export compatible Getters for matching service specific error codes and messages in the style above. See the Error Matchers heading in each respective library for details.

class AsError a where

Minimal complete definition

_Error

Methods

_Error :: Prism' a Error

A general Amazonka error.

_TransportError :: Prism' a HttpException

An error occured while communicating over HTTP with a remote service.

_SerializeError :: Prism' a SerializeError

A serialisation error occured when attempting to deserialise a response.

_ServiceError :: Prism' a ServiceError

A service specific error returned by the remote service.

class AsAuthError a where

Minimal complete definition

_AuthError

Methods

_AuthError :: Prism' a AuthError

A general authentication error.

_RetrievalError :: Prism' a HttpException

An error occured while communicating over HTTP with the local metadata endpoint.

_MissingEnvError :: Prism' a Text

The named environment variable was not found.

_InvalidEnvError :: Prism' a Text

An error occured parsing named environment variable's value.

_MissingFileError :: Prism' a FilePath

The specified credentials file could not be found.

_InvalidFileError :: Prism' a Text

An error occured parsing the credentials file.

_InvalidIAMError :: Prism' a Text

The specified IAM profile could not be found or deserialised.

trying :: MonadCatch m => Getting (First a) SomeException a -> m r -> m (Either a r)

A variant of try that takes a Prism (or any Fold) to select which exceptions are caught (c.f. tryJust, catchJust). If the Exception does not match the predicate, it is re-thrown.

trying :: MonadCatch m => Prism'     SomeException a -> m r -> m (Either a r)
trying :: MonadCatch m => Lens'      SomeException a -> m r -> m (Either a r)
trying :: MonadCatch m => Traversal' SomeException a -> m r -> m (Either a r)
trying :: MonadCatch m => Iso'       SomeException a -> m r -> m (Either a r)
trying :: MonadCatch m => Getter     SomeException a -> m r -> m (Either a r)
trying :: MonadCatch m => Fold       SomeException a -> m r -> m (Either a r)

catching :: MonadCatch m => Getting (First a) SomeException a -> m r -> (a -> m r) -> m r

Catch exceptions that match a given Prism (or any Fold, really).

>>> catching _AssertionFailed (assert False (return "uncaught")) $ \ _ -> return "caught"
"caught"

catching :: MonadCatch m => Prism' SomeException a     -> m r -> (a -> m r) -> m r
catching :: MonadCatch m => Lens' SomeException a      -> m r -> (a -> m r) -> m r
catching :: MonadCatch m => Traversal' SomeException a -> m r -> (a -> m r) -> m r
catching :: MonadCatch m => Iso' SomeException a       -> m r -> (a -> m r) -> m r
catching :: MonadCatch m => Getter SomeException a     -> m r -> (a -> m r) -> m r
catching :: MonadCatch m => Fold SomeException a       -> m r -> (a -> m r) -> m r

Building Error Prisms

_MatchServiceError :: AsError a => Service -> ErrorCode -> Getting (First ServiceError) a ServiceError

Provides a generalised prism for catching a specific service error identified by the opaque service abbreviation and error code.

This can be used if the generated error prisms provided by Network.AWS.ServiceName.Types do not cover all the thrown error codes. For example to define a new error prism:

{-# LANGUAGE OverloadedStrings #-}

import Network.AWS.S3 (ServiceError, s3)

_NoSuchBucketPolicy :: AsError a => Getting (First ServiceError) a ServiceError
_NoSuchBucketPolicy = _MatchServiceError s3 "NoSuchBucketPolicy"

With example usage being:

>>> import Control.Exception.Lens (trying)
>>> :t trying _NoSuchBucketPolicy
MonadCatch m => m a -> m (Either ServiceError a)

hasService :: (Applicative f, Choice p) => Service -> Optic' p f ServiceError ServiceError

hasStatus :: (Applicative f, Choice p) => Int -> Optic' p f ServiceError ServiceError

hasCode :: (Applicative f, Choice p) => ErrorCode -> Optic' p f ServiceError ServiceError

Logging

The exposed logging interface is a primitive Logger function which gets threaded through service calls and serialisation routines. This allows the library to output useful information and diagnostics.

The newLogger function can be used to construct a simple logger which writes output to a Handle, but in most production code you should probably consider using a more robust logging library such as tinylog or fast-logger.

data LogLevel

Constructors

Info
Info messages supplied by the user - this level is not emitted by the library.

Error
Error messages only.

Debug
Useful debug information + info + error levels.

Trace
Includes potentially sensitive signing metadata, and non-streaming response bodies.

type Logger = LogLevel -> ByteStringBuilder -> IO ()

A function threaded through various request and serialisation routines to log informational and debug messages.

Constructing a Logger

newLogger :: MonadIO m => LogLevel -> Handle -> m Logger

This is a primitive logger which can be used to log builds to a Handle.

Note: A more sophisticated logging library such as tinylog or fast-logger should be used in production code.

Re-exported Types

module Network.AWS.Core