Network.AWS.Auth

Description

Specify or retrieve your Amazon AWS credentials from the underlying host.

Defaults

accessKey

Arguments

:: ByteString
AWS_ACCESS_KEY

Default access key environment variable.

secretKey

Arguments

:: ByteString
AWS_SECRET_KEY

Default secret key environment variable.

Specifying credentials

fromKeys :: AccessKey -> SecretKey -> Auth

Explicit access and secret keys.

fromSession :: AccessKey -> SecretKey -> SecurityToken -> Auth

A session containing the access key, secret key, and a security token.

Retrieving credentials

data Credentials

Determines how authentication information is retrieved.

Constructors

FromKeys AccessKey SecretKey
Explicit access and secret keys. Note: you can achieve the same result purely by using fromKeys.

FromSession AccessKey SecretKey SecurityToken
A session containing the access key, secret key, and a security token. Note: you can achieve the same result purely by using fromSession.

FromProfile ByteString
An IAM Profile name to lookup from the local EC2 instance-data.

FromEnv ByteString ByteString
Environment variables to lookup for the access and secret keys.

Discover
Attempt to read the default access and secret keys from the environment, falling back to the first available IAM profile if they are not set.

Note: This attempts to resolve http://instance-data rather than directly retrieving http://169.254.169.254 for IAM profile information to ensure the dns lookup terminates promptly if not running on EC2.

getAuth :: (Functor m, MonadBase IO m) => Credentials -> ExceptT Error m Auth

Retrieve authentication information from the environment or instance-data.