Network.AWS.CloudFormation.V2010_05_15

Description

AWS CloudFormation gives developers and systems administrators an easy way to create and manage a collection of related AWS resources, provisioning and updating them in an orderly and predictable fashion. You can use AWS CloudFormation’s sample templates or create your own templates to describe the AWS resources, and any associated dependencies or runtime parameters, required to run your application. You don’t need to figure out the order for provisioning AWS services or the subtleties of making those dependencies work. CloudFormation takes care of this for you. After the AWS resources are deployed, you can modify and update them in a controlled and predictable way, in effect applying version control to your AWS infrastructure the same way you do with your software.

Documentation

data CancelUpdateStack

Constructors

CancelUpdateStack

Fields

_cusiStackName :: Text

The name or the unique identifier associated with the stack.

cusiStackName :: Iso' CancelUpdateStack Text

data CancelUpdateStackResponse

Constructors

CancelUpdateStackResponse

createStack

Arguments

:: Text
_csiStackName

-> CreateStack

Minimum specification for a CreateStack request.

data CreateStack

Constructors

CreateStack

Fields

_csiStackName :: Text

The name associated with the stack. The name must be unique within your AWS account. Must contain only alphanumeric characters (case sensitive) and start with an alpha character. Maximum length of the name is 255 characters.

_csiCapabilities :: [Capability]

The list of capabilities that you want to allow in the stack. If your template contains certain resources, you must specify the CAPABILITY_IAM value for this parameter; otherwise, this action returns an InsufficientCapabilities error. The following resources require you to specify the capabilities parameter: AWS::CloudFormation::Stack, AWS::IAM::AccessKey, AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, AWS::IAM::Role, AWS::IAM::User, and AWS::IAM::UserToGroupAddition.

_csiDisableRollback :: Maybe Bool

Set to true to disable rollback of the stack if stack creation failed. You can specify either DisableRollback or OnFailure, but not both. Default: false.

_csiNotificationARNs :: [Text]

The Simple Notification Service (SNS) topic ARNs to publish stack related events. You can find your SNS topic ARNs using the SNS console or your Command Line Interface (CLI).

_csiOnFailure :: Maybe OnFailure

Determines what action will be taken if stack creation fails. This must be one of: DO_NOTHING, ROLLBACK, or DELETE. You can specify either OnFailure or DisableRollback, but not both. Default: ROLLBACK.

_csiParameters :: [Parameter]

A list of Parameter structures that specify input parameters for the stack.

_csiStackPolicyBody :: Maybe Text

Structure containing the stack policy body. For more information, go to Prevent Updates to Stack Resources in the AWS CloudFormation User Guide. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

_csiStackPolicyURL :: Maybe Text

Location of a file containing the stack policy. The URL must point to a policy (max size: 16KB) located in an S3 bucket in the same region as the stack. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

_csiTags :: [Tag]

A set of user-defined Tags to associate with this stack, represented by key/value pairs. Tags defined for the stack are propagated to EC2 resources that are created as part of the stack. A maximum number of 10 tags can be specified.

_csiTemplateBody :: Maybe Text

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information, go to Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must specify either the TemplateBody or the TemplateURL parameter, but not both.

_csiTemplateURL :: Maybe Text

Location of file containing the template body. The URL must point to a template (max size: 307,200 bytes) located in an S3 bucket in the same region as the stack. For more information, go to the Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must specify either the TemplateBody or the TemplateURL parameter, but not both.

_csiTimeoutInMinutes :: Maybe Integer

The amount of time that can pass before the stack status becomes CREATE_FAILED; if DisableRollback is not set or is set to false, the stack will be rolled back.

csiTimeoutInMinutes :: Lens' CreateStack (Maybe Integer)

csiTemplateURL :: Lens' CreateStack (Maybe Text)

csiTemplateBody :: Lens' CreateStack (Maybe Text)

csiTags :: Lens' CreateStack [Tag]

csiStackPolicyURL :: Lens' CreateStack (Maybe Text)

csiStackPolicyBody :: Lens' CreateStack (Maybe Text)

csiStackName :: Lens' CreateStack Text

csiParameters :: Lens' CreateStack [Parameter]

csiOnFailure :: Lens' CreateStack (Maybe OnFailure)

csiNotificationARNs :: Lens' CreateStack [Text]

csiDisableRollback :: Lens' CreateStack (Maybe Bool)

csiCapabilities :: Lens' CreateStack [Capability]

data CreateStackResponse

Constructors

CreateStackResponse

Fields

_csoStackId :: Maybe Text

Unique identifier of the stack.

csoStackId :: Iso' CreateStackResponse (Maybe Text)

data DeleteStack

Constructors

DeleteStack

Fields

_dsiStackName :: Text

The name or the unique identifier associated with the stack.

dsiStackName :: Iso' DeleteStack Text

data DeleteStackResponse

Constructors

DeleteStackResponse

describeStackEvents

Arguments

:: Text
_dseiStackName

-> DescribeStackEvents

Minimum specification for a DescribeStackEvents request.

data DescribeStackEvents

Constructors

DescribeStackEvents

Fields

_dseiStackName :: Text

The name or the unique identifier associated with the stack, which are not always interchangeable: Running stacks: You can specify either the stack's name or its unique stack ID. Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

_dseiNextToken :: Maybe Text

String that identifies the start of the next list of events, if there is one. Default: There is no default value.

dseiStackName :: Lens' DescribeStackEvents Text

dseiNextToken :: Lens' DescribeStackEvents (Maybe Text)

data DescribeStackEventsResponse

Constructors

DescribeStackEventsResponse

Fields

_dseoNextToken :: Maybe Text

String that identifies the start of the next list of events, if there is one.

_dseoStackEvents :: [StackEvent]

A list of StackEvents structures.

dseoStackEvents :: Lens' DescribeStackEventsResponse [StackEvent]

dseoNextToken :: Lens' DescribeStackEventsResponse (Maybe Text)

data DescribeStackResource

Constructors

DescribeStackResource

Fields

_dsrjLogicalResourceId :: Text

The logical name of the resource as specified in the template. Default: There is no default value.

_dsrjStackName :: Text

The name or the unique identifier associated with the stack, which are not always interchangeable: Running stacks: You can specify either the stack's name or its unique stack ID. Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

dsrjStackName :: Lens' DescribeStackResource Text

dsrjLogicalResourceId :: Lens' DescribeStackResource Text

data DescribeStackResourceResponse

Constructors

DescribeStackResourceResponse

Fields

_dsrpStackResourceDetail :: Maybe StackResourceDetail

A StackResourceDetail structure containing the description of the specified resource in the specified stack.

dsrpStackResourceDetail :: Iso' DescribeStackResourceResponse (Maybe StackResourceDetail)

describeStackResources :: DescribeStackResources

Minimum specification for a DescribeStackResources request.

data DescribeStackResources

Constructors

DescribeStackResources

Fields

_dsriLogicalResourceId :: Maybe Text

The logical name of the resource as specified in the template. Default: There is no default value.

_dsriPhysicalResourceId :: Maybe Text

The name or unique identifier that corresponds to a physical instance ID of a resource supported by AWS CloudFormation. For example, for an Amazon Elastic Compute Cloud (EC2) instance, PhysicalResourceId corresponds to the InstanceId. You can pass the EC2 InstanceId to DescribeStackResources to find which stack the instance belongs to and what other resources are part of the stack. Required: Conditional. If you do not specify PhysicalResourceId, you must specify StackName. Default: There is no default value.

_dsriStackName :: Maybe Text

The name or the unique identifier associated with the stack, which are not always interchangeable: Running stacks: You can specify either the stack's name or its unique stack ID. Deleted stacks: You must specify the unique stack ID. Default: There is no default value. Required: Conditional. If you do not specify StackName, you must specify PhysicalResourceId.

dsriStackName :: Lens' DescribeStackResources (Maybe Text)

dsriPhysicalResourceId :: Lens' DescribeStackResources (Maybe Text)

dsriLogicalResourceId :: Lens' DescribeStackResources (Maybe Text)

data DescribeStackResourcesResponse

Constructors

DescribeStackResourcesResponse

Fields

_dsroStackResources :: [StackResource]

A list of StackResource structures.

dsroStackResources :: Iso' DescribeStackResourcesResponse [StackResource]

describeStacks :: DescribeStacks

Minimum specification for a DescribeStacks request.

data DescribeStacks

Constructors

DescribeStacks

Fields

_dsjNextToken :: Maybe Text

String that identifies the start of the next list of stacks, if there is one.

_dsjStackName :: Maybe Text

The name or the unique identifier associated with the stack, which are not always interchangeable: Running stacks: You can specify either the stack's name or its unique stack ID. Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

dsjStackName :: Lens' DescribeStacks (Maybe Text)

dsjNextToken :: Lens' DescribeStacks (Maybe Text)

data DescribeStacksResponse

Constructors

DescribeStacksResponse

Fields

_dsoNextToken :: Maybe Text

String that identifies the start of the next list of stacks, if there is one.

_dsoStacks :: [Stack]

A list of stack structures.

dsoStacks :: Lens' DescribeStacksResponse [Stack]

dsoNextToken :: Lens' DescribeStacksResponse (Maybe Text)

estimateTemplateCost :: EstimateTemplateCost

Minimum specification for a EstimateTemplateCost request.

data EstimateTemplateCost

Constructors

EstimateTemplateCost

Fields

_etciParameters :: [Parameter]

A list of Parameter structures that specify input parameters.

_etciTemplateBody :: Maybe Text

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. (For more information, go to Template Anatomy in the AWS CloudFormation User Guide.) Conditional: You must pass TemplateBody or TemplateURL. If both are passed, only TemplateBody is used.

_etciTemplateURL :: Maybe Text

Location of file containing the template body. The URL must point to a template located in an S3 bucket in the same region as the stack. For more information, go to Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must pass TemplateURL or TemplateBody. If both are passed, only TemplateBody is used.

etciTemplateURL :: Lens' EstimateTemplateCost (Maybe Text)

etciTemplateBody :: Lens' EstimateTemplateCost (Maybe Text)

etciParameters :: Lens' EstimateTemplateCost [Parameter]

data EstimateTemplateCostResponse

Constructors

EstimateTemplateCostResponse

Fields

_etcoUrl :: Maybe Text

An AWS Simple Monthly Calculator URL with a query string that describes the resources required to run the template.

etcoUrl :: Iso' EstimateTemplateCostResponse (Maybe Text)

data GetStackPolicy

Constructors

GetStackPolicy

Fields

_gspiStackName :: Text

The name or stack ID that is associated with the stack whose policy you want to get.

gspiStackName :: Iso' GetStackPolicy Text

data GetStackPolicyResponse

Constructors

GetStackPolicyResponse

Fields

_gspoStackPolicyBody :: Maybe Text

Structure containing the stack policy body. (For more information, go to Prevent Updates to Stack Resources in the AWS CloudFormation User Guide.).

gspoStackPolicyBody :: Iso' GetStackPolicyResponse (Maybe Text)

data GetTemplate

Constructors

GetTemplate

Fields

_gtiStackName :: Text

The name or the unique identifier associated with the stack, which are not always interchangeable: Running stacks: You can specify either the stack's name or its unique stack ID. Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

gtiStackName :: Iso' GetTemplate Text

data GetTemplateResponse

Constructors

GetTemplateResponse

Fields

_gtoTemplateBody :: Maybe Text

Structure containing the template body. (For more information, go to Template Anatomy in the AWS CloudFormation User Guide.).

gtoTemplateBody :: Iso' GetTemplateResponse (Maybe Text)

listStackResources

Arguments

:: Text
_lsriStackName

-> ListStackResources

Minimum specification for a ListStackResources request.

data ListStackResources

Constructors

ListStackResources

Fields

_lsriStackName :: Text

The name or the unique identifier associated with the stack, which are not always interchangeable: Running stacks: You can specify either the stack's name or its unique stack ID. Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

_lsriNextToken :: Maybe Text

String that identifies the start of the next list of stack resource summaries, if there is one. Default: There is no default value.

lsriStackName :: Lens' ListStackResources Text

lsriNextToken :: Lens' ListStackResources (Maybe Text)

data ListStackResourcesResponse

Constructors

ListStackResourcesResponse

Fields

_lsroNextToken :: Maybe Text

String that identifies the start of the next list of stack resources, if there is one.

_lsroStackResourceSummaries :: [StackResourceSummary]

A list of StackResourceSummary structures.

lsroStackResourceSummaries :: Lens' ListStackResourcesResponse [StackResourceSummary]

lsroNextToken :: Lens' ListStackResourcesResponse (Maybe Text)

listStacks :: ListStacks

Minimum specification for a ListStacks request.

data ListStacks

Constructors

ListStacks

Fields

_lsiNextToken :: Maybe Text

String that identifies the start of the next list of stacks, if there is one. Default: There is no default value.

_lsiStackStatusFilter :: [StackStatus]

Stack status to use as a filter. Specify one or more stack status codes to list only stacks with the specified status codes. For a complete list of stack status codes, see the StackStatus parameter of the Stack data type.

lsiStackStatusFilter :: Lens' ListStacks [StackStatus]

lsiNextToken :: Lens' ListStacks (Maybe Text)

data ListStacksResponse

Constructors

ListStacksResponse

Fields

_lsoNextToken :: Maybe Text

String that identifies the start of the next list of stacks, if there is one.

_lsoStackSummaries :: [StackSummary]

A list of StackSummary structures containing information about the specified stacks.

lsoStackSummaries :: Lens' ListStacksResponse [StackSummary]

lsoNextToken :: Lens' ListStacksResponse (Maybe Text)

setStackPolicy

Arguments

:: Text
_sspiStackName

-> SetStackPolicy

Minimum specification for a SetStackPolicy request.

data SetStackPolicy

Constructors

SetStackPolicy

Fields

_sspiStackName :: Text

The name or stack ID that you want to associate a policy with.

_sspiStackPolicyBody :: Maybe Text

Structure containing the stack policy body. For more information, go to Prevent Updates to Stack Resources in the AWS CloudFormation User Guide. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

_sspiStackPolicyURL :: Maybe Text

Location of a file containing the stack policy. The URL must point to a policy (max size: 16KB) located in an S3 bucket in the same region as the stack. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

sspiStackPolicyURL :: Lens' SetStackPolicy (Maybe Text)

sspiStackPolicyBody :: Lens' SetStackPolicy (Maybe Text)

sspiStackName :: Lens' SetStackPolicy Text

data SetStackPolicyResponse

Constructors

SetStackPolicyResponse

data CloudFormation

Supported version (2010-05-15) of the AWS CloudFormation service.

xmlOptions :: Tagged a XMLOptions

data Capability

Constructors

CapabilityCapabilityIam
CAPABILITY_IAM

data OnFailure

Determines what action will be taken if stack creation fails. This must be one of: DO_NOTHING, ROLLBACK, or DELETE. You can specify either OnFailure or DisableRollback, but not both. Default: ROLLBACK.

Constructors

OnFailureDelete
DELETE

OnFailureDoNothing
DO_NOTHING

OnFailureRollback
ROLLBACK

data ResourceStatus

Current status of the resource.

Constructors

ResourceStatusCreateComplete
CREATE_COMPLETE

ResourceStatusCreateFailed
CREATE_FAILED

ResourceStatusCreateInProgress
CREATE_IN_PROGRESS

ResourceStatusDeleteComplete
DELETE_COMPLETE

ResourceStatusDeleteFailed
DELETE_FAILED

ResourceStatusDeleteInProgress
DELETE_IN_PROGRESS

ResourceStatusUpdateComplete
UPDATE_COMPLETE

ResourceStatusUpdateFailed
UPDATE_FAILED

ResourceStatusUpdateInProgress
UPDATE_IN_PROGRESS

data StackStatus

Current status of the stack.

Constructors

StackStatusCreateComplete
CREATE_COMPLETE

StackStatusCreateFailed
CREATE_FAILED

StackStatusCreateInProgress
CREATE_IN_PROGRESS

StackStatusDeleteComplete
DELETE_COMPLETE

StackStatusDeleteFailed
DELETE_FAILED

StackStatusDeleteInProgress
DELETE_IN_PROGRESS

StackStatusRollbackComplete
ROLLBACK_COMPLETE

StackStatusRollbackFailed
ROLLBACK_FAILED

StackStatusRollbackInProgress
ROLLBACK_IN_PROGRESS

StackStatusUpdateComplete
UPDATE_COMPLETE

StackStatusUpdateCompleteCleanupInProgress
UPDATE_COMPLETE_CLEANUP_IN_PROGRESS

StackStatusUpdateInProgress
UPDATE_IN_PROGRESS

StackStatusUpdateRollbackComplete
UPDATE_ROLLBACK_COMPLETE

StackStatusUpdateRollbackCompleteCleanupInProgress
UPDATE_ROLLBACK_COMPLETE_CLEANUP_IN_PROGRESS

StackStatusUpdateRollbackFailed
UPDATE_ROLLBACK_FAILED

StackStatusUpdateRollbackInProgress
UPDATE_ROLLBACK_IN_PROGRESS

data Output

The Output data type.

Constructors

Output

Fields

_rOutputValue :: Maybe Text

The value associated with the output.

_rOutputKey :: Maybe Text

The key associated with the output.

_rDescription :: Maybe Text

User defined description associated with the output.

data Parameter

The Parameter data type.

Constructors

Parameter

Fields

_qParameterValue :: Maybe Text

The value associated with the parameter.

_qParameterKey :: Maybe Text

The key associated with the parameter.

_qUsePreviousValue :: Maybe Bool

During a stack update, use the existing parameter value that is being used for the stack.

data Stack

The Stack data type.

Constructors

Stack

Fields

_uCreationTime :: ISO8601

Time at which the stack was created.

_uStackStatus :: StackStatus

Current status of the stack.

_uDisableRollback :: Maybe Bool

Boolean to enable or disable rollback on stack creation failures: true: disable rollback false: enable rollback.

_uLastUpdatedTime :: Maybe ISO8601

The time the stack was last updated. This field will only be returned if the stack has been updated at least once.

_uNotificationARNs :: [Text]

SNS topic ARNs to which stack related events are published.

_uStackStatusReason :: Maybe Text

Success/failure message associated with the stack status.

_uOutputs :: [Output]

A list of output structures.

_uParameters :: [Parameter]

A list of Parameter structures.

_uStackId :: Maybe Text

Unique identifier of the stack.

_uDescription :: Maybe Text

User defined description associated with the stack.

_uCapabilities :: [Capability]

The capabilities allowed in the stack.

_uTags :: [Tag]

A list of Tags that specify cost allocation information for the stack.

_uTimeoutInMinutes :: Maybe Integer

The amount of time within which stack creation should complete.

_uStackName :: Text

The name associated with the stack.

data StackEvent

The StackEvent data type.

Constructors

StackEvent

Fields

_sfLogicalResourceId :: Maybe Text

The logical name of the resource specified in the template.

_sfPhysicalResourceId :: Maybe Text

The name or unique identifier associated with the physical instance of the resource.

_sfResourceType :: Maybe Text

Type of resource. (For more information, go to AWS Resource Types Reference in the AWS CloudFormation User Guide.).

_sfResourceStatusReason :: Maybe Text

Success/failure message associated with the resource.

_sfResourceProperties :: Maybe Text

BLOB of the properties used to create the resource.

_sfResourceStatus :: Maybe ResourceStatus

Current status of the resource.

_sfStackId :: Text

The unique ID name of the instance of the stack.

_sfTimestamp :: ISO8601

Time the status was updated.

_sfEventId :: Text

The unique ID of this event.

_sfStackName :: Text

The name associated with a stack.

data StackResource

The StackResource data type.

Constructors

StackResource

Fields

_sshLogicalResourceId :: Text

The logical name of the resource specified in the template.

_sshPhysicalResourceId :: Maybe Text

The name or unique identifier that corresponds to a physical instance ID of a resource supported by AWS CloudFormation.

_sshResourceType :: Text

Type of resource. (For more information, go to AWS Resource Types Reference in the AWS CloudFormation User Guide.).

_sshResourceStatusReason :: Maybe Text

Success/failure message associated with the resource.

_sshResourceStatus :: ResourceStatus

Current status of the resource.

_sshStackId :: Maybe Text

Unique identifier of the stack.

_sshDescription :: Maybe Text

User defined description associated with the resource.

_sshTimestamp :: ISO8601

Time the status was updated.

_sshStackName :: Maybe Text

The name associated with the stack.

data StackResourceDetail

A StackResourceDetail structure containing the description of the specified resource in the specified stack.

Constructors

StackResourceDetail

Fields

_srdLogicalResourceId :: Text

The logical name of the resource specified in the template.

_srdPhysicalResourceId :: Maybe Text

The name or unique identifier that corresponds to a physical instance ID of a resource supported by AWS CloudFormation.

_srdResourceType :: Text

Type of resource. ((For more information, go to AWS Resource Types Reference in the AWS CloudFormation User Guide.).

_srdResourceStatusReason :: Maybe Text

Success/failure message associated with the resource.

_srdResourceStatus :: ResourceStatus

Current status of the resource.

_srdMetadata :: Maybe Text

The JSON format content of the Metadata attribute declared for the resource. For more information, see Metadata Attribute in the AWS CloudFormation User Guide.

_srdStackId :: Maybe Text

Unique identifier of the stack.

_srdDescription :: Maybe Text

User defined description associated with the resource.

_srdLastUpdatedTimestamp :: ISO8601

Time the status was updated.

_srdStackName :: Maybe Text

The name associated with the stack.

data StackResourceSummary

Contains high-level information about the specified stack resource.

Constructors

StackResourceSummary

Fields

_srtLogicalResourceId :: Text

The logical name of the resource specified in the template.

_srtPhysicalResourceId :: Maybe Text

The name or unique identifier that corresponds to a physical instance ID of the resource.

_srtResourceType :: Text

Type of resource. (For more information, go to AWS Resource Types Reference in the AWS CloudFormation User Guide.).

_srtResourceStatusReason :: Maybe Text

Success/failure message associated with the resource.

_srtResourceStatus :: ResourceStatus

Current status of the resource.

_srtLastUpdatedTimestamp :: ISO8601

Time the status was updated.

data StackSummary

The StackSummary Data Type.

Constructors

StackSummary

Fields

_ssyCreationTime :: ISO8601

The time the stack was created.

_ssyStackStatus :: StackStatus

The current status of the stack.

_ssyLastUpdatedTime :: Maybe ISO8601

The time the stack was last updated. This field will only be returned if the stack has been updated at least once.

_ssyStackStatusReason :: Maybe Text

Success/Failure message associated with the stack status.

_ssyTemplateDescription :: Maybe Text

The template description of the template used to create the stack.

_ssyDeletionTime :: Maybe ISO8601

The time the stack was deleted.

_ssyStackId :: Maybe Text

Unique stack identifier.

_ssyStackName :: Text

The name associated with the stack.

data Tag

The Tag type is used by CreateStack in the Tags parameter. It allows you to specify a key/value pair that can be used to store information related to cost allocation for an AWS CloudFormation stack.

Constructors

Tag

Fields

_yValue :: Maybe Text

Required. A string containing the value for this tag. You can specify a maximum of 256 characters for a tag value.

_yKey :: Maybe Text

Required. A string used to identify this tag. You can specify a maximum of 128 characters for a tag key. Tags owned by Amazon Web Services (AWS) have the reserved prefix: aws:.

data TemplateParameter

The TemplateParameter data type.

Constructors

TemplateParameter

Fields

_tqParameterKey :: Maybe Text

The name associated with the parameter.

_tqDefaultValue :: Maybe Text

The default value associated with the parameter.

_tqNoEcho :: Maybe Bool

Flag indicating whether the parameter should be displayed as plain text in logs and UIs.

_tqDescription :: Maybe Text

User defined description associated with the parameter.

rOutputValue :: Lens' Output (Maybe Text)

rOutputKey :: Lens' Output (Maybe Text)

rDescription :: Lens' Output (Maybe Text)

qUsePreviousValue :: Lens' Parameter (Maybe Bool)

qParameterValue :: Lens' Parameter (Maybe Text)

qParameterKey :: Lens' Parameter (Maybe Text)

uTimeoutInMinutes :: Lens' Stack (Maybe Integer)

uTags :: Lens' Stack [Tag]

uStackStatusReason :: Lens' Stack (Maybe Text)

uStackStatus :: Lens' Stack StackStatus

uStackName :: Lens' Stack Text

uStackId :: Lens' Stack (Maybe Text)

uParameters :: Lens' Stack [Parameter]

uOutputs :: Lens' Stack [Output]

uNotificationARNs :: Lens' Stack [Text]

uLastUpdatedTime :: Lens' Stack (Maybe ISO8601)

uDisableRollback :: Lens' Stack (Maybe Bool)

uDescription :: Lens' Stack (Maybe Text)

uCreationTime :: Lens' Stack ISO8601

uCapabilities :: Lens' Stack [Capability]

sfTimestamp :: Lens' StackEvent ISO8601

sfStackName :: Lens' StackEvent Text

sfStackId :: Lens' StackEvent Text

sfResourceType :: Lens' StackEvent (Maybe Text)

sfResourceStatusReason :: Lens' StackEvent (Maybe Text)

sfResourceStatus :: Lens' StackEvent (Maybe ResourceStatus)

sfResourceProperties :: Lens' StackEvent (Maybe Text)

sfPhysicalResourceId :: Lens' StackEvent (Maybe Text)

sfLogicalResourceId :: Lens' StackEvent (Maybe Text)

sfEventId :: Lens' StackEvent Text

sshTimestamp :: Lens' StackResource ISO8601

sshStackName :: Lens' StackResource (Maybe Text)

sshStackId :: Lens' StackResource (Maybe Text)

sshResourceType :: Lens' StackResource Text

sshResourceStatusReason :: Lens' StackResource (Maybe Text)

sshResourceStatus :: Lens' StackResource ResourceStatus

sshPhysicalResourceId :: Lens' StackResource (Maybe Text)

sshLogicalResourceId :: Lens' StackResource Text

sshDescription :: Lens' StackResource (Maybe Text)

srdStackName :: Lens' StackResourceDetail (Maybe Text)

srdStackId :: Lens' StackResourceDetail (Maybe Text)

srdResourceType :: Lens' StackResourceDetail Text

srdResourceStatusReason :: Lens' StackResourceDetail (Maybe Text)

srdResourceStatus :: Lens' StackResourceDetail ResourceStatus

srdPhysicalResourceId :: Lens' StackResourceDetail (Maybe Text)

srdMetadata :: Lens' StackResourceDetail (Maybe Text)

srdLogicalResourceId :: Lens' StackResourceDetail Text

srdLastUpdatedTimestamp :: Lens' StackResourceDetail ISO8601

srdDescription :: Lens' StackResourceDetail (Maybe Text)

srtResourceType :: Lens' StackResourceSummary Text

srtResourceStatusReason :: Lens' StackResourceSummary (Maybe Text)

srtResourceStatus :: Lens' StackResourceSummary ResourceStatus

srtPhysicalResourceId :: Lens' StackResourceSummary (Maybe Text)

srtLogicalResourceId :: Lens' StackResourceSummary Text

srtLastUpdatedTimestamp :: Lens' StackResourceSummary ISO8601

ssyTemplateDescription :: Lens' StackSummary (Maybe Text)

ssyStackStatusReason :: Lens' StackSummary (Maybe Text)

ssyStackStatus :: Lens' StackSummary StackStatus

ssyStackName :: Lens' StackSummary Text

ssyStackId :: Lens' StackSummary (Maybe Text)

ssyLastUpdatedTime :: Lens' StackSummary (Maybe ISO8601)

ssyDeletionTime :: Lens' StackSummary (Maybe ISO8601)

ssyCreationTime :: Lens' StackSummary ISO8601

yValue :: Lens' Tag (Maybe Text)

yKey :: Lens' Tag (Maybe Text)

tqParameterKey :: Lens' TemplateParameter (Maybe Text)

tqNoEcho :: Lens' TemplateParameter (Maybe Bool)

tqDescription :: Lens' TemplateParameter (Maybe Text)

tqDefaultValue :: Lens' TemplateParameter (Maybe Text)

updateStack

Arguments

:: Text
_usiStackName

-> UpdateStack

Minimum specification for a UpdateStack request.

data UpdateStack

Constructors

UpdateStack

Fields

_usiStackName :: Text

The name or stack ID of the stack to update. Must contain only alphanumeric characters (case sensitive) and start with an alpha character. Maximum length of the name is 255 characters.

_usiCapabilities :: [Capability]

The list of capabilities that you want to allow in the stack. If your stack contains IAM resources, you must specify the CAPABILITY_IAM value for this parameter; otherwise, this action returns an InsufficientCapabilities error. IAM resources are the following: AWS::IAM::AccessKey, AWS::IAM::Group, AWS::IAM::Policy, AWS::IAM::User, and AWS::IAM::UserToGroupAddition.

_usiNotificationARNs :: [Text]

Update the ARNs for the Amazon SNS topics that are associated with the stack.

_usiParameters :: [Parameter]

A list of Parameter structures that specify input parameters for the stack.

_usiStackPolicyBody :: Maybe Text

Structure containing a new stack policy body. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both. You might update the stack policy, for example, in order to protect a new resource that you created during a stack update. If you do not specify a stack policy, the current policy that is associated with the stack is unchanged.

_usiStackPolicyDuringUpdateBody :: Maybe Text

Structure containing the temporary overriding stack policy body. You can specify either the StackPolicyDuringUpdateBody or the StackPolicyDuringUpdateURL parameter, but not both. If you want to update protected resources, specify a temporary overriding stack policy during this update. If you do not specify a stack policy, the current policy that is associated with the stack will be used.

_usiStackPolicyDuringUpdateURL :: Maybe Text

Location of a file containing the temporary overriding stack policy. The URL must point to a policy (max size: 16KB) located in an S3 bucket in the same region as the stack. You can specify either the StackPolicyDuringUpdateBody or the StackPolicyDuringUpdateURL parameter, but not both. If you want to update protected resources, specify a temporary overriding stack policy during this update. If you do not specify a stack policy, the current policy that is associated with the stack will be used.

_usiStackPolicyURL :: Maybe Text

Location of a file containing the updated stack policy. The URL must point to a policy (max size: 16KB) located in an S3 bucket in the same region as the stack. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both. You might update the stack policy, for example, in order to protect a new resource that you created during a stack update. If you do not specify a stack policy, the current policy that is associated with the stack is unchanged.

_usiTemplateBody :: Maybe Text

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. (For more information, go to Template Anatomy in the AWS CloudFormation User Guide.) Conditional: You must specify either the TemplateBody or the TemplateURL parameter, but not both.

_usiTemplateURL :: Maybe Text

Location of file containing the template body. The URL must point to a template located in an S3 bucket in the same region as the stack. For more information, go to Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must specify either the TemplateBody or the TemplateURL parameter, but not both.

_usiUsePreviousTemplate :: Maybe Bool

Reuse the existing template that is associated with the stack that you are updating.

usiUsePreviousTemplate :: Lens' UpdateStack (Maybe Bool)

usiTemplateURL :: Lens' UpdateStack (Maybe Text)

usiTemplateBody :: Lens' UpdateStack (Maybe Text)

usiStackPolicyURL :: Lens' UpdateStack (Maybe Text)

usiStackPolicyDuringUpdateURL :: Lens' UpdateStack (Maybe Text)

usiStackPolicyDuringUpdateBody :: Lens' UpdateStack (Maybe Text)

usiStackPolicyBody :: Lens' UpdateStack (Maybe Text)

usiStackName :: Lens' UpdateStack Text

usiParameters :: Lens' UpdateStack [Parameter]

usiNotificationARNs :: Lens' UpdateStack [Text]

usiCapabilities :: Lens' UpdateStack [Capability]

data UpdateStackResponse

Constructors

UpdateStackResponse

Fields

_usoStackId :: Maybe Text

Unique identifier of the stack.

usoStackId :: Iso' UpdateStackResponse (Maybe Text)

validateTemplate :: ValidateTemplate

Minimum specification for a ValidateTemplate request.

data ValidateTemplate

Constructors

ValidateTemplate

Fields

_vtiTemplateBody :: Maybe Text

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information, go to Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must pass TemplateURL or TemplateBody. If both are passed, only TemplateBody is used.

_vtiTemplateURL :: Maybe Text

Location of file containing the template body. The URL must point to a template (max size: 307,200 bytes) located in an S3 bucket in the same region as the stack. For more information, go to Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must pass TemplateURL or TemplateBody. If both are passed, only TemplateBody is used.

vtiTemplateURL :: Lens' ValidateTemplate (Maybe Text)

vtiTemplateBody :: Lens' ValidateTemplate (Maybe Text)

data ValidateTemplateResponse

Constructors

ValidateTemplateResponse

Fields

_vtoCapabilities :: [Capability]

The capabilities found within the template. Currently, CAPABILITY_IAM is the only capability detected. If your template contains IAM resources, you must specify the CAPABILITY_IAM value for this parameter when you use the CreateStack or UpdateStack actions with your template; otherwise, those actions return an InsufficientCapabilities error.

_vtoCapabilitiesReason :: Maybe Text

The capabilities reason found within the template.

_vtoDescription :: Maybe Text

The description found within the template.

_vtoParameters :: [TemplateParameter]

A list of TemplateParameter structures.

vtoParameters :: Lens' ValidateTemplateResponse [TemplateParameter]

vtoDescription :: Lens' ValidateTemplateResponse (Maybe Text)

vtoCapabilitiesReason :: Lens' ValidateTemplateResponse (Maybe Text)

vtoCapabilities :: Lens' ValidateTemplateResponse [Capability]