Network.AWS.Data

Description

Re-exports some of the underlying textual and byte serialisation mechanisms for convenience.

Many of the AWS identifiers like S3's ObjectVersionId or ETag, as well as any nullary sum types such as Region have ToText and ToByteString instances, making it convenient to use the type classes to convert a value to its textual representation.

Text

class FromText a where

Methods

parser :: Parser a

fromText :: FromText a => Text -> Either String a

fromTextError :: Text -> Parser a

takeLowerText :: Parser Text

class ToText a where

Methods

toText :: a -> Text

ByteString

class ToByteString a where

Minimal complete definition

Nothing

Methods

toBS :: a -> ByteString

Log Messages

class ToLog a where

Methods

build :: a -> Builder