Network.AWS.Data

Description

Re-exports some of the underlying textual and byte serialisation mechanisms for convenience.

Many of the AWS identifiers like S3's ObjectVersionId or ETag, as well as any nullary sum types such as Region have ToText and ToByteString instances, making it convenient to use the type classes to convert a value to its textual representation.

Text

class FromText a where

Minimal complete definition

parser

Methods

parser :: Parser a

fromText :: FromText a => Text -> Either String a

fromTextError :: Text -> Parser a

Fail parsing with a Text error.

Constrained to the actual attoparsec monad to avoid exposing fail usage directly.

takeLowerText :: Parser Text

class ToText a where

Minimal complete definition

toText

Methods

toText :: a -> Text

ByteString

class ToByteString a where

Methods

toBS :: a -> ByteString

Log Messages

class ToLog a where

Minimal complete definition

build

Methods

build :: a -> Builder

Convert a value to a loggable builder.