Network.AWS.Kinesis.V2013_12_02.Types

Description

Amazon Kinesis is a managed service that scales elastically for real-time processing of streaming big data. The service takes in large streams of data records that can then be consumed in real time by multiple data-processing applications that can be run on Amazon EC2 instances.

Documentation

data Kinesis

Supported version (2013-12-02) of the Amazon Kinesis service.

data ShardIteratorType

Determines how the shard iterator is used to start reading data records from the shard. The following are the valid shard iterator types: AT_SEQUENCE_NUMBER - Start reading exactly from the position denoted by a specific sequence number. AFTER_SEQUENCE_NUMBER - Start reading right after the position denoted by a specific sequence number. TRIM_HORIZON - Start reading at the last untrimmed record in the shard in the system, which is the oldest data record in the shard. LATEST - Start reading just after the most recent record in the shard, so that you always read the most recent data in the shard.

Constructors

ShardIteratorTypeAfterSequenceNumber
AFTER_SEQUENCE_NUMBER

ShardIteratorTypeAtSequenceNumber
AT_SEQUENCE_NUMBER

ShardIteratorTypeLatest
LATEST

ShardIteratorTypeTrimHorizon
TRIM_HORIZON

data StreamStatus

The current status of the stream being described. The stream status is one of the following states: CREATING - The stream is being created. Upon receiving a CreateStream request, Amazon Kinesis immediately returns and sets StreamStatus to CREATING. DELETING - The stream is being deleted. After a DeleteStream request, the specified stream is in the DELETING state until Amazon Kinesis completes the deletion. ACTIVE - The stream exists and is ready for read and write operations or deletion. You should perform read and write operations only on an ACTIVE stream. UPDATING - Shards in the stream are being merged or split. Read and write operations continue to work while the stream is in the UPDATING state.

Constructors

StreamStatusActive
ACTIVE

StreamStatusCreating
CREATING

StreamStatusDeleting
DELETING

StreamStatusUpdating
UPDATING

data HashKeyRange

The range of possible hash key values for the shard, which is a set of ordered contiguous positive integers.

Constructors

HashKeyRange

Fields

_hkrEndingHashKey :: Text

The ending hash key of the hash key range.

_hkrStartingHashKey :: Text

The starting hash key of the hash key range.

data Record

The unit of data of the Amazon Kinesis stream, which is composed of a sequence number, a partition key, and a data blob.

Constructors

Record

Fields

_rSequenceNumber :: Text

The unique identifier for the record in the Amazon Kinesis stream.

_rPartitionKey :: Text

Identifies which shard in the stream the data record is assigned to.

_rData :: Base64

The data blob. The data in the blob is both opaque and immutable to the Amazon Kinesis service, which does not inspect, interpret, or change the data in the blob in any way. The maximum size of the data blob (the payload after Base64-decoding) is 50 kilobytes (KB).

data SequenceNumberRange

The range of possible sequence numbers for the shard.

Constructors

SequenceNumberRange

Fields

_snrStartingSequenceNumber :: Text

The starting sequence number for the range.

_snrEndingSequenceNumber :: Maybe Text

The ending sequence number for the range. Shards that are in the OPEN state have an ending sequence number of null.

data Shard

A uniquely identified group of data records in an Amazon Kinesis stream.

Constructors

Shard

Fields

_sAdjacentParentShardId :: Maybe Text

The shard Id of the shard adjacent to the shard's parent.

_sHashKeyRange :: HashKeyRange

The range of possible hash key values for the shard, which is a set of ordered contiguous positive integers.

_sParentShardId :: Maybe Text

The shard Id of the shard's parent.

_sSequenceNumberRange :: SequenceNumberRange

The range of possible sequence numbers for the shard.

_sShardId :: Text

The unique identifier of the shard within the Amazon Kinesis stream.

data StreamDescription

Contains the current status of the stream, the stream ARN, an array of shard objects that comprise the stream, and states whether there are more shards available.

Constructors

StreamDescription

Fields

_sdStreamStatus :: StreamStatus

The current status of the stream being described. The stream status is one of the following states: CREATING - The stream is being created. Upon receiving a CreateStream request, Amazon Kinesis immediately returns and sets StreamStatus to CREATING. DELETING - The stream is being deleted. After a DeleteStream request, the specified stream is in the DELETING state until Amazon Kinesis completes the deletion. ACTIVE - The stream exists and is ready for read and write operations or deletion. You should perform read and write operations only on an ACTIVE stream. UPDATING - Shards in the stream are being merged or split. Read and write operations continue to work while the stream is in the UPDATING state.

_sdHasMoreShards :: Bool

If set to true there are more shards in the stream available to describe.

_sdStreamARN :: Text

The Amazon Resource Name (ARN) for the stream being described.

_sdShards :: [Shard]

The shards that comprise the stream.

_sdStreamName :: Text

The name of the stream being described.

hkrStartingHashKey :: Lens' HashKeyRange Text

hkrEndingHashKey :: Lens' HashKeyRange Text

rSequenceNumber :: Lens' Record Text

rPartitionKey :: Lens' Record Text

rData :: Lens' Record Base64

snrStartingSequenceNumber :: Lens' SequenceNumberRange Text

snrEndingSequenceNumber :: Lens' SequenceNumberRange (Maybe Text)

sShardId :: Lens' Shard Text

sSequenceNumberRange :: Lens' Shard SequenceNumberRange

sParentShardId :: Lens' Shard (Maybe Text)

sHashKeyRange :: Lens' Shard HashKeyRange

sAdjacentParentShardId :: Lens' Shard (Maybe Text)

sdStreamStatus :: Lens' StreamDescription StreamStatus

sdStreamName :: Lens' StreamDescription Text

sdStreamARN :: Lens' StreamDescription Text

sdShards :: Lens' StreamDescription [Shard]

sdHasMoreShards :: Lens' StreamDescription Bool