Network.AWS.Presign

Description

This module contains functions for presigning requests using MonadIO and not one of the AWS specific transformers.

It is intended for use directly with Network.AWS.Auth when only presigning and no other AWS actions are required.

Documentation

presignURL

Arguments

:: (MonadIO m, AWSRequest a)

=> Auth

-> Region

-> UTCTime
Signing time.

-> Seconds
Expiry time.

-> a
Request to presign.

-> m ByteString

Presign an URL that is valid from the specified time until the number of seconds expiry has elapsed.

See: presign, presignWith

presign

Arguments

:: (MonadIO m, AWSRequest a)

=> Auth

-> Region

-> UTCTime
Signing time.

-> Seconds
Expiry time.

-> a
Request to presign.

-> m ClientRequest

Presign an HTTP request that is valid from the specified time until the number of seconds expiry has elapsed.

See: presignWith

presignWith

Arguments

:: (MonadIO m, AWSRequest a)

=> (Service -> Service)
Modify the default service configuration.

-> Auth

-> Region

-> UTCTime
Signing time.

-> Seconds
Expiry time.

-> a
Request to presign.

-> m ClientRequest

A variant of presign that allows modifying the default Service definition used to configure the request.