Network.AWS.S3

Description

Amazon Simple Storage Service (Amazon S3) is a web service that enables you to store data in the cloud.

You can then download the data or use the data with other AWS services, such as Amazon Elastic Compute Cloud (Amazon EC2).

PUT Object

data PutObject

Add an object to a bucket.

You must have WRITE permissions on a bucket to add an object to it.

Amazon S3 never adds partial objects; if you receive a success response, Amazon S3 added the entire object to the bucket.

Amazon S3 is a distributed system. If it receives multiple write requests for the same object simultaneously, it overwrites all but the last object written.

Amazon S3 does not provide object locking; if you need this, make sure to build it into your application layer or use versioning instead.

If you enable versioning for a bucket, Amazon S3 automatically generates a unique version ID for the object being stored. Amazon S3 returns this ID in the response using the x-amz-version-id response header.

If versioning is suspended, Amazon S3 always uses null as the version ID for the object stored.

http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectPUT.html

Constructors

PutObject

Fields

poBucket :: !Bucket

poKey :: !Key

poHeaders :: [AnyHeader]

poBody :: Body

Data Types

s3 :: Service

Currently supported version of the S3 service.

data S3ErrorResponse

Constructors

S3ErrorResponse

Fields

sssError :: !Text

newtype Key

Constructors

Key Text

newtype Bucket

Constructors

Bucket Text

newtype S3HeadersResponse

Constructors

S3HeadersResponse [(ByteString, ByteString)]

data AES256

Constructors

AES256

data CannedACL

Constructors

Private

PublicRead

PublicReadWrite

AuthenticatedRead

BucketOwnerRead

BucketOwnerFullControl

LogDeliveryWrite

data StorageClass

Constructors

Standard

ReducedRedundancy

type Storage = Header "x-amz-storage-class" StorageClass

type RedirectLocation = Header "x-amz-website-redirect-location" Text

type Encryption = Header "x-amz-server-side-encryption" AES256

type ACL = Header "x-amz-acl" CannedACL

type GrantRead = Header "x-amz-grant-read" Text

type GrantWrite = Header "x-amz-grant-write" Text

type GrantReadACP = Header "x-amz-grant-read-acp" Text

type GrantWriteACP = Header "x-amz-grant-write-acp" Text

type GrantFullControl = Header "x-amz-grant-full-control" Text

type Metadata = Header "x-amz-meta-" (Text, Text)

s3NS :: ByteString

XML namespace to annotate S3 elements with.

s3Elem :: ByteString -> NName ByteString

Helper to define S3 namespaced XML elements.

Common

module Network.AWS