Network.AWS.S3

Description

Amazon Simple Storage Service (Amazon S3) is a web service that enables you to store data in the cloud.

You can then download the data or use the data with other AWS services, such as Amazon Elastic Compute Cloud (Amazon EC2).

Operations on the Service

GET Service

data GetService

Returns a list of all buckets owned by the requester.

http://docs.aws.amazon.com/AmazonS3/latest/API/RESTServiceGET.html

Constructors

GetService

data GetServiceResponse

Constructors

GetServiceResponse

Fields

gsrOwner :: !Owner

Information about the bucket owner.

gsrBuckets :: [Bucket]

A list of buckets for the service.

Operations on Objects

DELETE Object

data DeleteObject

Removes the null version (if there is one) of an object and inserts a delete marker, which becomes the latest version of the object.

If there isn't a null version, Amazon S3 does not remove any objects.

http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectDELETE.html

Constructors

DeleteObject

Fields

doBucket :: !Text

doKey :: !Text

doHeaders :: [AnyHeader]

newtype DeleteObjectResponse

Constructors

DeleteObjectResponse [(ByteString, ByteString)]

POST Delete Multiple Objects

data DeleteMultipleObjects

Delete multiple objects from a bucket using a single HTTP request.

If you know the object keys that you want to delete, then this operation provides a suitable alternative to sending individual delete requests (see DeleteObject), reducing per-request overhead.

http://docs.aws.amazon.com/AmazonS3/latest/API/multiobjectdeleteapi.html

Constructors

DeleteMultipleObjects

Fields

dmoBucket :: !Text

dmoMD5 :: !MD5

dmoHeaders :: [AnyHeader]

dmoObjects :: !DMObjects

data DeleteMultipleObjectsResponse

Constructors

DeleteMultipleObjectsResponse

Fields

dmorDeleted :: [DeletedObject]

dmorError :: [DeleteError]

GET Object

data GetObject

Retrieves an object from Amazon S3.

You must have READ access to the object.

If you grant READ access to the anonymous user, you can return the object without using an authorization header.

http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectGET.html

Constructors

GetObject

Fields

goBucket :: !Text

goKey :: !Text

goHeaders :: [AnyHeader]

type GetObjectResponse = S3BodyResponse

GET Object ACL

data GetObjectACL

Uses the ACL subresource to return the access control list (ACL) of an object.

You must have READ_ACP access to the object.

By default, returns ACL information about the latest version of an object. To return ACL information about a different version, use the versionId subresource.

http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectGETacl.html

Constructors

GetObjectACL

Fields

goaclBucket :: !Text

goaclKey :: !Text

goaclHeaders :: [AnyHeader]

data GetObjectACLResponse

Constructors

GetObjectACLResponse

Fields

goaclrHeaders :: [(ByteString, ByteString)]

goaclrPolicy :: !AccessControlPolicy

GET Object Torrent

data GetObjectTorrent

Use the torrent subresource to return torrent files from a bucket.

You can get torrent only for objects that are less than 5 GB in size.

http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectGETtorrent.html

Constructors

GetObjectTorrent

Fields

gotBucket :: !Text

gotKey :: !Text

gotHeaders :: [AnyHeader]

type GetObjectTorrentResponse = S3BodyResponse

HEAD Object

data HeadObject

Retrieves metadata from an object without returning the object itself.

You must have READ access to the object.

By default, the HEAD operation retrieves metadata from the latest version of an object. If the latest version is a delete marker, Amazon S3 behaves as if the object was deleted. To retrieve metadata from a different version, use the versionId subresource.

http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectHEAD.html

Constructors

HeadObject

Fields

hoBucket :: !Text

hoKey :: !Text

hoHeaders :: [AnyHeader]

type HeadObjectResponse = S3HeaderResponse

OPTIONS Object

data OptionsObject

Preflight request to determine if an actual request can be sent with the specific origin, HTTP method, and headers.

Amazon S3 supports cross-origin resource sharing (CORS) by enabling you to add a cors subresource on a bucket.

When a browser sends this preflight request, Amazon S3 responds by evaluating the rules that are defined in the cors configuration.

If cors is not enabled on the bucket, then Amazon S3 returns a 403 Forbidden response.

http://docs.aws.amazon.com/AmazonS3/latest/API/RESTOPTIONSobject.html

Constructors

OptionsObject

Fields

ooBucket :: !Text

ooKey :: !Text

ooOrigin :: !Origin

ooRequestMethod :: !AccessControlRequestMethod

ooRequestHeaders :: !AccessControlRequestHeaders

ooHeaders :: [AnyHeader]

type OptionsObjectResponse = S3HeaderResponse

POST Object Restore

data PostObjectRestore

Restores a temporary copy of an archived object.

In the request, you specify the number of days that you want the restored copy to exist. After the specified period, Amazon S3 deletes the temporary copy.

Note that the object remains archived; Amazon S3 deletes only the restored copy.

An object in the Glacier storage class is an archived object. To access the object, you must first initiate a restore request, which restores a copy of the archived object. Restore jobs typically complete in three to five hours.

http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectPOSTrestore.html

Constructors

PostObjectRestore

Fields

porBucket :: !Text

porKey :: !Text

porMD5 :: !MD5

porRestoreRequest :: !RestoreRequest

type PostObjectRestoreResponse = S3HeaderResponse

PUT Object

data PutObject

Add an object to a bucket.

You must have WRITE permissions on a bucket to add an object to it.

Amazon S3 never adds partial objects; if you receive a success response, Amazon S3 added the entire object to the bucket.

Amazon S3 is a distributed system. If it receives multiple write requests for the same object simultaneously, it overwrites all but the last object written.

Amazon S3 does not provide object locking; if you need this, make sure to build it into your application layer or use versioning instead.

If you enable versioning for a bucket, Amazon S3 automatically generates a unique version ID for the object being stored. Amazon S3 returns this ID in the response using the x-amz-version-id response header.

If versioning is suspended, Amazon S3 always uses null as the version ID for the object stored.

http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectPUT.html

Constructors

PutObject

Fields

poBucket :: !Text

poKey :: !Text

poHeaders :: [AnyHeader]

poBody :: Body

type PutObjectResponse = S3HeaderResponse

PUT Object ACL

data PutObjectACL

Set the access control list (ACL) permissions for an object that already exists in a bucket.

You must have WRITE_ACP permission to set the ACL of an object.

You can use one of the following two ways to set an object's permissions: * Specify the ACL in the request body, or * Specify permissions using request headers

Depending on your application needs, you may choose to set the ACL on an object using either the request body or the headers.

For example, if you have an existing application that updates an object ACL using the request body, then you can continue to use that approach.

http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectPUTacl.html

Constructors

PutObjectACL

Fields

poaclBucket :: !Text

poaclKey :: !Text

poaclPolicy :: Maybe AccessControlPolicy

poaclHeaders :: [AnyHeader]

type PutObjectACLResponse = S3HeaderResponse

PUT Object Copy

data PutObjectCopy

Create a copy of an object that is already stored in Amazon S3.

A copy operation is the same as performing a GET and then a PUT.

Note You can store individual objects of up to 5 TB in Amazon S3.

You create a copy of your object up to 5 GB in size in a single atomic operation using this API. However, for copying an object greater than 5 GB, you must use the multipart upload API.

When copying an object, you can preserve most of the metadata (default) or specify new metadata. However, the ACL is not preserved and is set to private for the user making the request.

By default, x-amz-copy-source identifies the latest version of an object to copy. (If the latest version is a Delete Marker, Amazon S3 behaves as if the object was deleted.) To copy a different version, use the versionId subresource.

If you enable Versioning on the target bucket, Amazon S3 generates a unique version ID for the object being copied. This version ID is different from the version ID of the source object.

http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectCOPY.html

Constructors

PutObjectCopy

Fields

pocBucket :: !Text

pocKey :: !Text

pocSource :: !CopySource

pocHeaders :: [AnyHeader]

type PutObjectCopyResponse = S3HeaderResponse

POST Initiate Multipart Upload

data InitiateMultipartUpload

Initiate a multipart upload and return an upload ID.

This upload ID is used to associate all the parts in the specific multipart upload. You specify this upload ID in each of your subsequent upload part requests (see UploadPart).

You also include this upload ID in the final request to either complete or abort the multipart upload request.

http://docs.aws.amazon.com/AmazonS3/latest/API/mpUploadInitiate.html

Constructors

InitiateMultipartUpload

Fields

imuBucket :: !Text

imuKey :: !Text

imuHeaders :: [AnyHeader]

data InitiateMultipartUploadResponse

PUT Upload Part

data UploadPart

Upload a part in a multipart upload.

Note In this operation you provide part data in your request. However, you have an option to specify your existing Amazon S3 object as data source for the part your are uploading.

To upload a part from an existing object you use the UploadPartCopy operation.

You must initiate a multipart upload (see InitiateMultipartUpload) before you can upload any part.

In response to your initiate request. Amazon S3 returns an upload ID, a unique identifier, that you must include in your upload part request.

Part numbers can be any number from 1 to 10,000, inclusive. A part number uniquely identifies a part and also defines its position within the object being created.

If you upload a new part using the same part number that was used with a previous part, the previously uploaded part is overwritten.

Each part must be at least 5 MB in size, except the last part.

There is no size limit on the last part of your multipart upload.

http://docs.aws.amazon.com/AmazonS3/latest/API/mpUploadUploadPart.html

Constructors

UploadPart

Fields

upBucket :: !Text

upKey :: !Text

upPartNumber :: !Text

upUploadId :: !Text

upHeaders :: [AnyHeader]

upBody :: !Body

PUT Upload Part Copy

data UploadPartCopy

Uploads a part by copying data from an existing object as data source.

You specify the data source by adding the request header x-amz-copy-source in your request and a byte range by adding the request header x-amz-copy-source-range in your request.

Note Instead of using an existing object as part data, you might use the UploadPart operation and provide data in your request. For more information, see Upload Part.

http://docs.aws.amazon.com/AmazonS3/latest/API/mpUploadUploadPartCopy.html

Constructors

UploadPartCopy

Fields

upcBucket :: !Text

upcKey :: !Text

upcSource :: !CopySource

upcPartNumber :: !Text

upcUploadId :: !Text

upcHeaders :: [AnyHeader]

POST Complete Multipart Upload

data CompleteMultipartUpload

Completes a multipart upload by assembling previously uploaded parts.

You first initiate the multipart upload and then upload all parts using the UploadPart operation.

After successfully uploading all relevant parts of an upload, you call this operation to complete the upload.

Upon receiving this request, Amazon S3 concatenates all the parts in ascending order by part number to create a new object.

In the CompleteMultipartUpload request, you must provide the parts list.

You must ensure the parts list is complete, this operation concatenates the parts you provide in the list.

For each part in the list, you must provide the part number and the ETag header value, returned after that part was uploaded.

Processing of a Complete Multipart Upload request could take several minutes to complete. After Amazon S3 begins processing the request, it sends an HTTP response header that specifies a 200 OK response.

While processing is in progress, Amazon S3 periodically sends whitespace characters to keep the connection from timing out. Because a request could fail after the initial 200 OK response has been sent, it is important that you check the response body to determine whether the request succeeded.

http://docs.aws.amazon.com/AmazonS3/latest/API/mpUploadComplete.html

Constructors

CompleteMultipartUpload

Fields

cmuBucket :: !Text

cmuKey :: !Text

cmuUploadId :: !Text

cmuHeaders :: [AnyHeader]

cmuParts :: [Part]

DELETE Abort Multipart Upload

data AbortMultipartUpload

Aborts a multipart upload.

After a multipart upload is aborted, no additional parts can be uploaded using that upload ID.

The storage consumed by any previously uploaded parts will be freed.

However, if any part uploads are currently in progress, those part uploads might or might not succeed. As a result, it might be necessary to abort a given multipart upload multiple times in order to completely free all storage consumed by all parts.

http://docs.aws.amazon.com/AmazonS3/latest/API/mpUploadAbort.html

Constructors

AbortMultipartUpload

Fields

amuBucket :: !Text

amuKey :: !Text

amuUploadId :: !Text

GET List Parts

data ListParts

List the parts that have been uploaded for a specific multipart upload.

This operation must include the upload ID, which you obtain by sending InitiateMultipartUpload.

This returns a maximum of 1,000 uploaded parts and the default number of parts returned is 1,000 parts.

You can restrict the number of parts returned by specifying the maxParts parameter.

If your multipart upload consists of more than 1,000 parts, the response returns an IsTruncated field with the value of True, and a NextPartNumberMarker element.

In subsequent List Parts requests you can include the part-number-marker query string parameter and set its value to the NextPartNumberMarker field value from the previous response.

http://docs.aws.amazon.com/AmazonS3/latest/API/mpUploadListParts.html

Constructors

ListParts

Data Types

s3 :: Service

Currently supported version of the S3 service.

newtype S3HeaderResponse

Constructors

S3HeaderResponse [(ByteString, ByteString)]

data S3BodyResponse

Constructors

S3BodyResponse

Fields

s3Headers :: [(ByteString, ByteString)]

s3Body :: InputStream ByteString

data S3ErrorResponse

Constructors

S3ErrorResponse

Fields

sssError :: !Text

newtype S3HeadersResponse

Constructors

S3HeadersResponse [(ByteString, ByteString)]

data Bucket

Constructors

Bucket

Fields

bName :: !Text

Bucket's name.

bCreationDate :: !UTCTime

Date the bucket was created.

data DMObject

Constructors

DMObject

Fields

oKey :: !Text

oVersion :: Maybe Text

data DMObjects

Constructors

DMObjects

Fields

oQuiet :: !Bool

oObjects :: [DMObject]

data DeletedObject

Constructors

DeletedObject

Fields

dobjKey :: !Text

dobjVersionId :: Maybe Text

dobjDeleteMarker :: Maybe Bool

dobjDeleteMarkerVersionId :: Maybe Text

data DeleteError

Constructors

DeleteError

Fields

deKey :: !Text

deVersionId :: Maybe Text

deCode :: !Text

deMessage :: !Text

data Owner

Constructors

Owner

Fields

oID :: !Text

Bucket owner's user ID.

oDisplayName :: !Text

Bucket owner's display name.

data Grantee

Constructors

Grantee

Fields

gID :: !Text

gDisplayName :: !Text

data Grant

Constructors

Grant

Fields

gGrantee :: !Grantee

gPermission :: !Text

data AccessControlList

Constructors

AccessControlList

Fields

aclGrant :: [Grant]

data AccessControlPolicy

Constructors

AccessControlPolicy

Fields

acpOwner :: !Owner

acpAccessControlList :: !AccessControlList

data RestoreRequest

Constructors

RestoreRequest

Fields

rrDays :: !Int

data Part

Constructors

Part

Fields

pPartNumber :: !Text

pETag :: !ETag

data AES256

Constructors

AES256

data CannedACL

Constructors

Private

PublicRead

PublicReadWrite

AuthenticatedRead

BucketOwnerRead

BucketOwnerFullControl

LogDeliveryWrite

data StorageClass

Constructors

Standard

ReducedRedundancy

data Directive

Constructors

Copy

Replace

data Source

Constructors

Source

Fields

srcBucket :: !Text

srcKey :: !Text

newtype ETag

Constructors

ETag

Fields

unETag :: Text

type Metadata = Header "x-amz-meta-" (Text, Text)

type Storage = Header "x-amz-storage-class" StorageClass

type ACL = Header "x-amz-acl" CannedACL

type RedirectLocation = Header "x-amz-website-redirect-location" Text

When a bucket is configured as a website, you can set this metadata on the object so the website endpoint will evaluate the request for the object as a 301 redirect to another object in the same bucket or an external URL.

type Encryption = Header "x-amz-server-side-encryption" AES256

If the object is stored using server-side encryption, response includes this header with value of the encryption algorithm used.

type ResponseContentType = Header "response-content-type" Text

Sets the Content-Type header of the response.

type ResponseContentLanguage = Header "response-content-language" Text

Sets the Content-Language header of the response.

type ResponseExpires = Header "response-expires" Text

Sets the Expires header of the response.

type ResponseCacheControl = Header "response-cache-control" Text

Sets the Cache-Control header of the response.

type ResponseContentDisposition = Header "response-content-disposition" Text

Sets the Content-Disposition header of the response.

type ResponseContentEncoding = Header "response-content-encoding" Text

Sets the Content-Encoding header of the response.

type DeleteMarker = Header "x-amz-delete-marker" Bool

Specifies whether the object retrieved was (true) or was not (false) a Delete Marker.

If false, this response header does not appear in the response.

type Expiration = Header "x-amz-expiration" UTCTime

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key value pairs providing object expiration information. The value of the rule-id is URL encoded.

type Restore = Header "x-amz-restore" Text

Provides information about object restoration operation and expiration time of the restored object copy.

type VersionId = Header "x-amz-version-id" Text

Returns the version ID of the retrieved object if it has a unique version ID.

type MFA = Header "x-amz-mfa" Text

The value is the concatenation of the authentication device's serial number, a space, and the value displayed on your authentication device.

Condition: Required to permanently delete a versioned object if versioning is configured with MFA Delete enabled.

type Origin = Header "origin" Text

Identifies the origin of the cross-origin request to Amazon S3.

For example, http:www.example.com.

type AccessControlRequestMethod = Header "access-control-request-method" Method

Identifies what HTTP method will be used in the actual request.

type AccessControlRequestHeaders = Header "access-control-request-headers" [Text]

A comma-delimited list of HTTP headers that will be sent in the actual request.

type CopySource = Header "x-amz-copy-source" Source

The name of the source bucket and key name of the source object, separated by a slash (/).

This string must be URL-encoded. Additionally, the source bucket must be valid and you must have READ access to the valid source object.

If the source object is archived in Amazon Glacier (storage class of the object is GLACIER), you must first restore a temporary copy using the POST Object restore. Otherwise, Amazon S3 returns the 403 ObjectNotInActiveTierError error response.

type CopySourceRange = Header "x-amz-copy-source-range" Text

The range of bytes to copy from the source object.

The range value must use the form bytes=first-last, where the first and last are the zero-based byte offsets to copy.

For example, bytes=0-9 indicates that you want to copy the first ten bytes of the source.

type MetadataDirective = Header "x-amz-metadata-directive" Directive

Specifies whether the metadata is copied from the source object or replaced with metadata provided in the request.

If copied, the metadata, except for the version ID, remains unchanged. In addition, the server-side-encryption, storage-class, and website-redirect-location metadata from the source is not copied. If you specify this metadata explicitly in the copy request, Amazon S3 adds this metadata to the resulting object. If you specify headers in the request specifying any user-defined metadata, Amazon S3 ignores these headers.

If replaced, all original metadata is replaced by the metadata you specify.

type CopySourceIfMatch = Header "x-amz-copy-source-if-match" ETag

Copies the object if its entity tag (ETag) matches the specified tag; otherwise, the request returns a 412 HTTP status code error.

Constraints: This header can be used with x-amz-copy-source-if-unmodified-since, but cannot be used with other conditional copy headers.

type CopySourceIfNoneMatch = Header "x-amz-copy-source-if-none-match" ETag

Copies the object if its entity tag (ETag) is different than the specified ETag; otherwise, the request returns a 412 HTTP status code error.

Constraints: This header can be used with x-amz-copy-source-if-modified-since, but cannot be used with other conditional copy headers.

type CopySourceIfUnmodifiedSince = Header "x-amz-copy-source-if-unmodified-since" UTCTime

Copies the object if it hasn't been modified since the specified time; otherwise, the request returns a 412 HTTP status code error.

type CopySourceIfModifiedSince = Header "x-amz-copy-source-if-modified-since" UTCTime

Copies the object if it has been modified since the specified time; otherwise, the request returns a 412 HTTP status code error.

s3NS :: ByteString

XML namespace to annotate S3 elements with.

s3Elem :: ByteString -> NName ByteString

Helper to define S3 namespaced XML elements.

Common

module Network.AWS