Network.AWS.S3.V2006_03_01.CompleteMultipartUpload

Description

Completes a multipart upload by assembling previously uploaded parts.

Documentation

data CompleteMultipartUpload

Constructors

CompleteMultipartUpload

Fields

_cmusBucket :: BucketName

_cmusUploadId :: Text

_cmusKey :: ObjectKey

_cmusMultipartUpload :: CompletedMultipartUpload

data CompleteMultipartUploadResponse

Constructors

CompleteMultipartUploadResponse

Fields

_cmupBucket :: Maybe BucketName

_cmupETag :: Maybe ETag

Entity tag of the object.

_cmupLocation :: Maybe Text

_cmupKey :: Maybe ObjectKey

_cmupExpiration :: Maybe RFC822

If the object expiration is configured, this will contain the expiration date (expiry-date) and rule ID (rule-id). The value of rule-id is URL encoded.

_cmupVersionId :: Maybe ObjectVersionId

Version of the object.

_cmupServerSideEncryption :: Maybe ServerSideEncryption

The Server-side encryption algorithm used when storing this object in S3.